Amazonka.AMP

Description

Derived from API version 2020-08-01 of the AWS service descriptions, licensed under Apache 2.0.

Amazon Managed Service for Prometheus

Service Configuration

defaultService :: Service

API version 2020-08-01 of the Amazon Prometheus Service SDK configuration.

Errors

Error matchers are designed for use with the functions provided by Control.Exception.Lens. This allows catching (and rethrowing) service specific errors returned by AMP.

ValidationException

_ValidationException :: AsError a => Getting (First ServiceError) a ServiceError

The input fails to satisfy the constraints specified by an AWS service.

AccessDeniedException

_AccessDeniedException :: AsError a => Getting (First ServiceError) a ServiceError

User does not have sufficient access to perform this action.

ConflictException

_ConflictException :: AsError a => Getting (First ServiceError) a ServiceError

Updating or deleting a resource can cause an inconsistent state.

ServiceQuotaExceededException

_ServiceQuotaExceededException :: AsError a => Getting (First ServiceError) a ServiceError

Request would cause a service quota to be exceeded.

ThrottlingException

_ThrottlingException :: AsError a => Getting (First ServiceError) a ServiceError

Request was denied due to request throttling.

InternalServerException

_InternalServerException :: AsError a => Getting (First ServiceError) a ServiceError

Unexpected error during processing of request.

ResourceNotFoundException

_ResourceNotFoundException :: AsError a => Getting (First ServiceError) a ServiceError

Request references a resource which does not exist.

Waiters

Waiters poll by repeatedly sending a request until some remote success condition configured by the Wait specification is fulfilled. The Wait specification determines how many attempts should be made, in addition to delay and retry strategies.

WorkspaceDeleted

newWorkspaceDeleted :: Wait DescribeWorkspace

Polls DescribeWorkspace every 2 seconds until a successful state is reached. An error is returned after 60 failed checks.

WorkspaceActive

newWorkspaceActive :: Wait DescribeWorkspace

Polls DescribeWorkspace every 2 seconds until a successful state is reached. An error is returned after 60 failed checks.

Operations

Some AWS operations return results that are incomplete and require subsequent requests in order to obtain the entire result set. The process of sending subsequent requests to continue where a previous request left off is called pagination. For example, the ListObjects operation of Amazon S3 returns up to 1000 objects at a time, and you must send subsequent requests with the appropriate Marker in order to retrieve the next page of results.

Operations that have an AWSPager instance can transparently perform subsequent requests, correctly setting Markers and other request facets to iterate through the entire result set of a truncated API operation. Operations which support this have an additional note in the documentation.

Many operations have the ability to filter results on the server side. See the individual operation parameters for details.

ListWorkspaces (Paginated)

data ListWorkspaces

Represents the input of a ListWorkspaces operation.

See: newListWorkspaces smart constructor.

Constructors

ListWorkspaces' (Maybe Text) (Maybe Text) (Maybe Natural)

newListWorkspaces :: ListWorkspaces

Create a value of ListWorkspaces with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:alias:ListWorkspaces', listWorkspaces_alias - Optional filter for workspace alias. Only the workspaces with aliases that begin with this value will be returned.

$sel:nextToken:ListWorkspaces', listWorkspaces_nextToken - Pagination token to request the next page in a paginated list. This token is obtained from the output of the previous ListWorkspaces request.

$sel:maxResults:ListWorkspaces', listWorkspaces_maxResults - Maximum results to return in response (default=100, maximum=1000).

data ListWorkspacesResponse

Represents the output of a ListWorkspaces operation.

See: newListWorkspacesResponse smart constructor.

Constructors

ListWorkspacesResponse' (Maybe Text) Int [WorkspaceSummary]

newListWorkspacesResponse

Arguments

:: Int
$sel:httpStatus:ListWorkspacesResponse'

-> ListWorkspacesResponse

Create a value of ListWorkspacesResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:nextToken:ListWorkspaces', listWorkspacesResponse_nextToken - Pagination token to use when requesting the next page in this list.

$sel:httpStatus:ListWorkspacesResponse', listWorkspacesResponse_httpStatus - The response's http status code.

$sel:workspaces:ListWorkspacesResponse', listWorkspacesResponse_workspaces - The list of existing workspaces, including those undergoing creation or deletion.

CreateAlertManagerDefinition

data CreateAlertManagerDefinition

Represents the input of a CreateAlertManagerDefinition operation.

See: newCreateAlertManagerDefinition smart constructor.

Constructors

CreateAlertManagerDefinition' (Maybe Text) Base64 Text

newCreateAlertManagerDefinition

Arguments

:: ByteString
$sel:data':CreateAlertManagerDefinition'

-> Text
$sel:workspaceId:CreateAlertManagerDefinition'

-> CreateAlertManagerDefinition

Create a value of CreateAlertManagerDefinition with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:clientToken:CreateAlertManagerDefinition', createAlertManagerDefinition_clientToken - Optional, unique, case-sensitive, user-provided identifier to ensure the idempotency of the request.

$sel:data':CreateAlertManagerDefinition', createAlertManagerDefinition_data - The alert manager definition data.-- -- Note: This Lens automatically encodes and decodes Base64 data. -- The underlying isomorphism will encode to Base64 representation during -- serialisation, and decode from Base64 representation during deserialisation. -- This Lens accepts and returns only raw unencoded data.

$sel:workspaceId:CreateAlertManagerDefinition', createAlertManagerDefinition_workspaceId - The ID of the workspace in which to create the alert manager definition.

data CreateAlertManagerDefinitionResponse

Represents the output of a CreateAlertManagerDefinition operation.

See: newCreateAlertManagerDefinitionResponse smart constructor.

Constructors

CreateAlertManagerDefinitionResponse' Int AlertManagerDefinitionStatus

newCreateAlertManagerDefinitionResponse

Arguments

:: Int
$sel:httpStatus:CreateAlertManagerDefinitionResponse'

-> AlertManagerDefinitionStatus
$sel:status:CreateAlertManagerDefinitionResponse'

-> CreateAlertManagerDefinitionResponse

Create a value of CreateAlertManagerDefinitionResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:httpStatus:CreateAlertManagerDefinitionResponse', createAlertManagerDefinitionResponse_httpStatus - The response's http status code.

$sel:status:CreateAlertManagerDefinitionResponse', createAlertManagerDefinitionResponse_status - The status of alert manager definition.

DeleteWorkspace

data DeleteWorkspace

Represents the input of a DeleteWorkspace operation.

See: newDeleteWorkspace smart constructor.

Constructors

DeleteWorkspace' (Maybe Text) Text

newDeleteWorkspace

Arguments

:: Text
$sel:workspaceId:DeleteWorkspace'

-> DeleteWorkspace

Create a value of DeleteWorkspace with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:clientToken:DeleteWorkspace', deleteWorkspace_clientToken - Optional, unique, case-sensitive, user-provided identifier to ensure the idempotency of the request.

$sel:workspaceId:DeleteWorkspace', deleteWorkspace_workspaceId - The ID of the workspace to delete.

data DeleteWorkspaceResponse

See: newDeleteWorkspaceResponse smart constructor.

Constructors

DeleteWorkspaceResponse'

newDeleteWorkspaceResponse :: DeleteWorkspaceResponse

Create a value of DeleteWorkspaceResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

ListTagsForResource

data ListTagsForResource

See: newListTagsForResource smart constructor.

Constructors

ListTagsForResource' Text

newListTagsForResource

Arguments

:: Text
$sel:resourceArn:ListTagsForResource'

-> ListTagsForResource

Create a value of ListTagsForResource with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:resourceArn:ListTagsForResource', listTagsForResource_resourceArn - The ARN of the resource.

data ListTagsForResourceResponse

See: newListTagsForResourceResponse smart constructor.

Constructors

ListTagsForResourceResponse' (Maybe (HashMap Text Text)) Int

newListTagsForResourceResponse

Arguments

:: Int
$sel:httpStatus:ListTagsForResourceResponse'

-> ListTagsForResourceResponse

Create a value of ListTagsForResourceResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:tags:ListTagsForResourceResponse', listTagsForResourceResponse_tags - Undocumented member.

$sel:httpStatus:ListTagsForResourceResponse', listTagsForResourceResponse_httpStatus - The response's http status code.

CreateRuleGroupsNamespace

data CreateRuleGroupsNamespace

Represents the input of a CreateRuleGroupsNamespace operation.

See: newCreateRuleGroupsNamespace smart constructor.

Constructors

CreateRuleGroupsNamespace' (Maybe Text) (Maybe (HashMap Text Text)) Base64 Text Text

newCreateRuleGroupsNamespace

Arguments

:: ByteString
$sel:data':CreateRuleGroupsNamespace'

-> Text
$sel:name:CreateRuleGroupsNamespace'

-> Text
$sel:workspaceId:CreateRuleGroupsNamespace'

-> CreateRuleGroupsNamespace

Create a value of CreateRuleGroupsNamespace with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:clientToken:CreateRuleGroupsNamespace', createRuleGroupsNamespace_clientToken - Optional, unique, case-sensitive, user-provided identifier to ensure the idempotency of the request.

$sel:tags:CreateRuleGroupsNamespace', createRuleGroupsNamespace_tags - Optional, user-provided tags for this rule groups namespace.

$sel:data':CreateRuleGroupsNamespace', createRuleGroupsNamespace_data - The namespace data that define the rule groups.-- -- Note: This Lens automatically encodes and decodes Base64 data. -- The underlying isomorphism will encode to Base64 representation during -- serialisation, and decode from Base64 representation during deserialisation. -- This Lens accepts and returns only raw unencoded data.

$sel:name:CreateRuleGroupsNamespace', createRuleGroupsNamespace_name - The rule groups namespace name.

$sel:workspaceId:CreateRuleGroupsNamespace', createRuleGroupsNamespace_workspaceId - The ID of the workspace in which to create the rule group namespace.

data CreateRuleGroupsNamespaceResponse

Represents the output of a CreateRuleGroupsNamespace operation.

See: newCreateRuleGroupsNamespaceResponse smart constructor.

Constructors

CreateRuleGroupsNamespaceResponse' (Maybe (HashMap Text Text)) Int Text Text RuleGroupsNamespaceStatus

newCreateRuleGroupsNamespaceResponse

Arguments

:: Int
$sel:httpStatus:CreateRuleGroupsNamespaceResponse'

-> Text
$sel:arn:CreateRuleGroupsNamespaceResponse'

-> Text
$sel:name:CreateRuleGroupsNamespace'

-> RuleGroupsNamespaceStatus
$sel:status:CreateRuleGroupsNamespaceResponse'

-> CreateRuleGroupsNamespaceResponse

Create a value of CreateRuleGroupsNamespaceResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:tags:CreateRuleGroupsNamespace', createRuleGroupsNamespaceResponse_tags - The tags of this rule groups namespace.

$sel:httpStatus:CreateRuleGroupsNamespaceResponse', createRuleGroupsNamespaceResponse_httpStatus - The response's http status code.

$sel:arn:CreateRuleGroupsNamespaceResponse', createRuleGroupsNamespaceResponse_arn - The Amazon Resource Name (ARN) of this rule groups namespace.

$sel:name:CreateRuleGroupsNamespace', createRuleGroupsNamespaceResponse_name - The rule groups namespace name.

$sel:status:CreateRuleGroupsNamespaceResponse', createRuleGroupsNamespaceResponse_status - The status of rule groups namespace.

DescribeAlertManagerDefinition

data DescribeAlertManagerDefinition

Represents the input of a DescribeAlertManagerDefinition operation.

See: newDescribeAlertManagerDefinition smart constructor.

Constructors

DescribeAlertManagerDefinition' Text

newDescribeAlertManagerDefinition

Arguments

:: Text
$sel:workspaceId:DescribeAlertManagerDefinition'

-> DescribeAlertManagerDefinition

Create a value of DescribeAlertManagerDefinition with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:workspaceId:DescribeAlertManagerDefinition', describeAlertManagerDefinition_workspaceId - The ID of the workspace to describe.

data DescribeAlertManagerDefinitionResponse

Represents the output of a DescribeAlertManagerDefinition operation.

See: newDescribeAlertManagerDefinitionResponse smart constructor.

Constructors

DescribeAlertManagerDefinitionResponse' Int AlertManagerDefinitionDescription

newDescribeAlertManagerDefinitionResponse

Arguments

:: Int
$sel:httpStatus:DescribeAlertManagerDefinitionResponse'

-> AlertManagerDefinitionDescription
$sel:alertManagerDefinition:DescribeAlertManagerDefinitionResponse'

-> DescribeAlertManagerDefinitionResponse

Create a value of DescribeAlertManagerDefinitionResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:httpStatus:DescribeAlertManagerDefinitionResponse', describeAlertManagerDefinitionResponse_httpStatus - The response's http status code.

$sel:alertManagerDefinition:DescribeAlertManagerDefinitionResponse', describeAlertManagerDefinitionResponse_alertManagerDefinition - The properties of the selected workspace's alert manager definition.

DescribeWorkspace

data DescribeWorkspace

Represents the input of a DescribeWorkspace operation.

See: newDescribeWorkspace smart constructor.

Constructors

DescribeWorkspace' Text

newDescribeWorkspace

Arguments

:: Text
$sel:workspaceId:DescribeWorkspace'

-> DescribeWorkspace

Create a value of DescribeWorkspace with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:workspaceId:DescribeWorkspace', describeWorkspace_workspaceId - The ID of the workspace to describe.

data DescribeWorkspaceResponse

Represents the output of a DescribeWorkspace operation.

See: newDescribeWorkspaceResponse smart constructor.

Constructors

DescribeWorkspaceResponse' Int WorkspaceDescription

newDescribeWorkspaceResponse

Arguments

:: Int
$sel:httpStatus:DescribeWorkspaceResponse'

-> WorkspaceDescription
$sel:workspace:DescribeWorkspaceResponse'

-> DescribeWorkspaceResponse

Create a value of DescribeWorkspaceResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:httpStatus:DescribeWorkspaceResponse', describeWorkspaceResponse_httpStatus - The response's http status code.

$sel:workspace:DescribeWorkspaceResponse', describeWorkspaceResponse_workspace - The properties of the selected workspace.

PutAlertManagerDefinition

data PutAlertManagerDefinition

Represents the input of a PutAlertManagerDefinition operation.

See: newPutAlertManagerDefinition smart constructor.

Constructors

PutAlertManagerDefinition' (Maybe Text) Base64 Text

newPutAlertManagerDefinition

Arguments

:: ByteString
$sel:data':PutAlertManagerDefinition'

-> Text
$sel:workspaceId:PutAlertManagerDefinition'

-> PutAlertManagerDefinition

Create a value of PutAlertManagerDefinition with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:clientToken:PutAlertManagerDefinition', putAlertManagerDefinition_clientToken - Optional, unique, case-sensitive, user-provided identifier to ensure the idempotency of the request.

$sel:data':PutAlertManagerDefinition', putAlertManagerDefinition_data - The alert manager definition data.-- -- Note: This Lens automatically encodes and decodes Base64 data. -- The underlying isomorphism will encode to Base64 representation during -- serialisation, and decode from Base64 representation during deserialisation. -- This Lens accepts and returns only raw unencoded data.

$sel:workspaceId:PutAlertManagerDefinition', putAlertManagerDefinition_workspaceId - The ID of the workspace in which to update the alert manager definition.

data PutAlertManagerDefinitionResponse

Represents the output of a PutAlertManagerDefinition operation.

See: newPutAlertManagerDefinitionResponse smart constructor.

Constructors

PutAlertManagerDefinitionResponse' Int AlertManagerDefinitionStatus

newPutAlertManagerDefinitionResponse

Arguments

:: Int
$sel:httpStatus:PutAlertManagerDefinitionResponse'

-> AlertManagerDefinitionStatus
$sel:status:PutAlertManagerDefinitionResponse'

-> PutAlertManagerDefinitionResponse

Create a value of PutAlertManagerDefinitionResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:httpStatus:PutAlertManagerDefinitionResponse', putAlertManagerDefinitionResponse_httpStatus - The response's http status code.

$sel:status:PutAlertManagerDefinitionResponse', putAlertManagerDefinitionResponse_status - The status of alert manager definition.

DeleteAlertManagerDefinition

data DeleteAlertManagerDefinition

Represents the input of a DeleteAlertManagerDefinition operation.

See: newDeleteAlertManagerDefinition smart constructor.

Constructors

DeleteAlertManagerDefinition' (Maybe Text) Text

newDeleteAlertManagerDefinition

Arguments

:: Text
$sel:workspaceId:DeleteAlertManagerDefinition'

-> DeleteAlertManagerDefinition

Create a value of DeleteAlertManagerDefinition with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:clientToken:DeleteAlertManagerDefinition', deleteAlertManagerDefinition_clientToken - Optional, unique, case-sensitive, user-provided identifier to ensure the idempotency of the request.

$sel:workspaceId:DeleteAlertManagerDefinition', deleteAlertManagerDefinition_workspaceId - The ID of the workspace in which to delete the alert manager definition.

data DeleteAlertManagerDefinitionResponse

See: newDeleteAlertManagerDefinitionResponse smart constructor.

Constructors

DeleteAlertManagerDefinitionResponse'

newDeleteAlertManagerDefinitionResponse :: DeleteAlertManagerDefinitionResponse

Create a value of DeleteAlertManagerDefinitionResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

DescribeRuleGroupsNamespace

data DescribeRuleGroupsNamespace

Represents the input of a DescribeRuleGroupsNamespace operation.

See: newDescribeRuleGroupsNamespace smart constructor.

Constructors

DescribeRuleGroupsNamespace' Text Text

newDescribeRuleGroupsNamespace

Arguments

:: Text
$sel:name:DescribeRuleGroupsNamespace'

-> Text
$sel:workspaceId:DescribeRuleGroupsNamespace'

-> DescribeRuleGroupsNamespace

Create a value of DescribeRuleGroupsNamespace with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:name:DescribeRuleGroupsNamespace', describeRuleGroupsNamespace_name - The rule groups namespace.

$sel:workspaceId:DescribeRuleGroupsNamespace', describeRuleGroupsNamespace_workspaceId - The ID of the workspace to describe.

data DescribeRuleGroupsNamespaceResponse

Represents the output of a DescribeRuleGroupsNamespace operation.

See: newDescribeRuleGroupsNamespaceResponse smart constructor.

Constructors

DescribeRuleGroupsNamespaceResponse' Int RuleGroupsNamespaceDescription

newDescribeRuleGroupsNamespaceResponse

Arguments

:: Int
$sel:httpStatus:DescribeRuleGroupsNamespaceResponse'

-> RuleGroupsNamespaceDescription
$sel:ruleGroupsNamespace:DescribeRuleGroupsNamespaceResponse'

-> DescribeRuleGroupsNamespaceResponse

Create a value of DescribeRuleGroupsNamespaceResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:httpStatus:DescribeRuleGroupsNamespaceResponse', describeRuleGroupsNamespaceResponse_httpStatus - The response's http status code.

$sel:ruleGroupsNamespace:DescribeRuleGroupsNamespaceResponse', describeRuleGroupsNamespaceResponse_ruleGroupsNamespace - The selected rule groups namespace.

UpdateWorkspaceAlias

data UpdateWorkspaceAlias

Represents the input of an UpdateWorkspaceAlias operation.

See: newUpdateWorkspaceAlias smart constructor.

Constructors

UpdateWorkspaceAlias' (Maybe Text) (Maybe Text) Text

newUpdateWorkspaceAlias

Arguments

:: Text
$sel:workspaceId:UpdateWorkspaceAlias'

-> UpdateWorkspaceAlias

Create a value of UpdateWorkspaceAlias with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:clientToken:UpdateWorkspaceAlias', updateWorkspaceAlias_clientToken - Optional, unique, case-sensitive, user-provided identifier to ensure the idempotency of the request.

$sel:alias:UpdateWorkspaceAlias', updateWorkspaceAlias_alias - The new alias of the workspace.

$sel:workspaceId:UpdateWorkspaceAlias', updateWorkspaceAlias_workspaceId - The ID of the workspace being updated.

data UpdateWorkspaceAliasResponse

See: newUpdateWorkspaceAliasResponse smart constructor.

Constructors

UpdateWorkspaceAliasResponse'

newUpdateWorkspaceAliasResponse :: UpdateWorkspaceAliasResponse

Create a value of UpdateWorkspaceAliasResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

DeleteRuleGroupsNamespace

data DeleteRuleGroupsNamespace

Represents the input of a DeleteRuleGroupsNamespace operation.

See: newDeleteRuleGroupsNamespace smart constructor.

Constructors

DeleteRuleGroupsNamespace' (Maybe Text) Text Text

newDeleteRuleGroupsNamespace

Arguments

:: Text
$sel:name:DeleteRuleGroupsNamespace'

-> Text
$sel:workspaceId:DeleteRuleGroupsNamespace'

-> DeleteRuleGroupsNamespace

Create a value of DeleteRuleGroupsNamespace with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:clientToken:DeleteRuleGroupsNamespace', deleteRuleGroupsNamespace_clientToken - Optional, unique, case-sensitive, user-provided identifier to ensure the idempotency of the request.

$sel:name:DeleteRuleGroupsNamespace', deleteRuleGroupsNamespace_name - The rule groups namespace name.

$sel:workspaceId:DeleteRuleGroupsNamespace', deleteRuleGroupsNamespace_workspaceId - The ID of the workspace to delete rule group definition.

data DeleteRuleGroupsNamespaceResponse

See: newDeleteRuleGroupsNamespaceResponse smart constructor.

Constructors

DeleteRuleGroupsNamespaceResponse'

newDeleteRuleGroupsNamespaceResponse :: DeleteRuleGroupsNamespaceResponse

Create a value of DeleteRuleGroupsNamespaceResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

PutRuleGroupsNamespace

data PutRuleGroupsNamespace

Represents the input of a PutRuleGroupsNamespace operation.

See: newPutRuleGroupsNamespace smart constructor.

Constructors

PutRuleGroupsNamespace' (Maybe Text) Base64 Text Text

newPutRuleGroupsNamespace

Arguments

:: ByteString
$sel:data':PutRuleGroupsNamespace'

-> Text
$sel:name:PutRuleGroupsNamespace'

-> Text
$sel:workspaceId:PutRuleGroupsNamespace'

-> PutRuleGroupsNamespace

Create a value of PutRuleGroupsNamespace with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:clientToken:PutRuleGroupsNamespace', putRuleGroupsNamespace_clientToken - Optional, unique, case-sensitive, user-provided identifier to ensure the idempotency of the request.

$sel:data':PutRuleGroupsNamespace', putRuleGroupsNamespace_data - The namespace data that define the rule groups.-- -- Note: This Lens automatically encodes and decodes Base64 data. -- The underlying isomorphism will encode to Base64 representation during -- serialisation, and decode from Base64 representation during deserialisation. -- This Lens accepts and returns only raw unencoded data.

$sel:name:PutRuleGroupsNamespace', putRuleGroupsNamespace_name - The rule groups namespace name.

$sel:workspaceId:PutRuleGroupsNamespace', putRuleGroupsNamespace_workspaceId - The ID of the workspace in which to update the rule group namespace.

data PutRuleGroupsNamespaceResponse

Represents the output of a PutRuleGroupsNamespace operation.

See: newPutRuleGroupsNamespaceResponse smart constructor.

Constructors

PutRuleGroupsNamespaceResponse' (Maybe (HashMap Text Text)) Int Text Text RuleGroupsNamespaceStatus

newPutRuleGroupsNamespaceResponse

Arguments

:: Int
$sel:httpStatus:PutRuleGroupsNamespaceResponse'

-> Text
$sel:arn:PutRuleGroupsNamespaceResponse'

-> Text
$sel:name:PutRuleGroupsNamespace'

-> RuleGroupsNamespaceStatus
$sel:status:PutRuleGroupsNamespaceResponse'

-> PutRuleGroupsNamespaceResponse

Create a value of PutRuleGroupsNamespaceResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:tags:PutRuleGroupsNamespaceResponse', putRuleGroupsNamespaceResponse_tags - The tags of this rule groups namespace.

$sel:httpStatus:PutRuleGroupsNamespaceResponse', putRuleGroupsNamespaceResponse_httpStatus - The response's http status code.

$sel:arn:PutRuleGroupsNamespaceResponse', putRuleGroupsNamespaceResponse_arn - The Amazon Resource Name (ARN) of this rule groups namespace.

$sel:name:PutRuleGroupsNamespace', putRuleGroupsNamespaceResponse_name - The rule groups namespace name.

$sel:status:PutRuleGroupsNamespaceResponse', putRuleGroupsNamespaceResponse_status - The status of rule groups namespace.

ListRuleGroupsNamespaces (Paginated)

data ListRuleGroupsNamespaces

Represents the input of a ListRuleGroupsNamespaces operation.

See: newListRuleGroupsNamespaces smart constructor.

Constructors

ListRuleGroupsNamespaces' (Maybe Text) (Maybe Text) (Maybe Natural) Text

newListRuleGroupsNamespaces

Arguments

:: Text
$sel:workspaceId:ListRuleGroupsNamespaces'

-> ListRuleGroupsNamespaces

Create a value of ListRuleGroupsNamespaces with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:nextToken:ListRuleGroupsNamespaces', listRuleGroupsNamespaces_nextToken - Pagination token to request the next page in a paginated list. This token is obtained from the output of the previous ListRuleGroupsNamespaces request.

$sel:name:ListRuleGroupsNamespaces', listRuleGroupsNamespaces_name - Optional filter for rule groups namespace name. Only the rule groups namespace that begin with this value will be returned.

$sel:maxResults:ListRuleGroupsNamespaces', listRuleGroupsNamespaces_maxResults - Maximum results to return in response (default=100, maximum=1000).

$sel:workspaceId:ListRuleGroupsNamespaces', listRuleGroupsNamespaces_workspaceId - The ID of the workspace.

data ListRuleGroupsNamespacesResponse

Represents the output of a ListRuleGroupsNamespaces operation.

See: newListRuleGroupsNamespacesResponse smart constructor.

Constructors

ListRuleGroupsNamespacesResponse' (Maybe Text) Int [RuleGroupsNamespaceSummary]

newListRuleGroupsNamespacesResponse

Arguments

:: Int
$sel:httpStatus:ListRuleGroupsNamespacesResponse'

-> ListRuleGroupsNamespacesResponse

Create a value of ListRuleGroupsNamespacesResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:nextToken:ListRuleGroupsNamespaces', listRuleGroupsNamespacesResponse_nextToken - Pagination token to use when requesting the next page in this list.

$sel:httpStatus:ListRuleGroupsNamespacesResponse', listRuleGroupsNamespacesResponse_httpStatus - The response's http status code.

$sel:ruleGroupsNamespaces:ListRuleGroupsNamespacesResponse', listRuleGroupsNamespacesResponse_ruleGroupsNamespaces - The list of the selected rule groups namespaces.

TagResource

data TagResource

See: newTagResource smart constructor.

Constructors

TagResource' Text (HashMap Text Text)

newTagResource

Arguments

:: Text
$sel:resourceArn:TagResource'

-> TagResource

Create a value of TagResource with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:resourceArn:TagResource', tagResource_resourceArn - The ARN of the resource.

$sel:tags:TagResource', tagResource_tags - Undocumented member.

data TagResourceResponse

See: newTagResourceResponse smart constructor.

Constructors

TagResourceResponse' Int

newTagResourceResponse

Arguments

:: Int
$sel:httpStatus:TagResourceResponse'

-> TagResourceResponse

Create a value of TagResourceResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:httpStatus:TagResourceResponse', tagResourceResponse_httpStatus - The response's http status code.

UntagResource

data UntagResource

See: newUntagResource smart constructor.

Constructors

UntagResource' Text [Text]

newUntagResource

Arguments

:: Text
$sel:resourceArn:UntagResource'

-> UntagResource

Create a value of UntagResource with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:resourceArn:UntagResource', untagResource_resourceArn - The ARN of the resource.

$sel:tagKeys:UntagResource', untagResource_tagKeys - One or more tag keys

data UntagResourceResponse

See: newUntagResourceResponse smart constructor.

Constructors

UntagResourceResponse' Int

newUntagResourceResponse

Arguments

:: Int
$sel:httpStatus:UntagResourceResponse'

-> UntagResourceResponse

Create a value of UntagResourceResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:httpStatus:UntagResourceResponse', untagResourceResponse_httpStatus - The response's http status code.

CreateWorkspace

data CreateWorkspace

Represents the input of a CreateWorkspace operation.

See: newCreateWorkspace smart constructor.

Constructors

CreateWorkspace' (Maybe Text) (Maybe Text) (Maybe (HashMap Text Text))

newCreateWorkspace :: CreateWorkspace

Create a value of CreateWorkspace with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:clientToken:CreateWorkspace', createWorkspace_clientToken - Optional, unique, case-sensitive, user-provided identifier to ensure the idempotency of the request.

$sel:alias:CreateWorkspace', createWorkspace_alias - An optional user-assigned alias for this workspace. This alias is for user reference and does not need to be unique.

$sel:tags:CreateWorkspace', createWorkspace_tags - Optional, user-provided tags for this workspace.

data CreateWorkspaceResponse

Represents the output of a CreateWorkspace operation.

See: newCreateWorkspaceResponse smart constructor.

Constructors

CreateWorkspaceResponse' (Maybe (HashMap Text Text)) Int Text WorkspaceStatus Text

newCreateWorkspaceResponse

Arguments

:: Int
$sel:httpStatus:CreateWorkspaceResponse'

-> Text
$sel:arn:CreateWorkspaceResponse'

-> WorkspaceStatus
$sel:status:CreateWorkspaceResponse'

-> Text
$sel:workspaceId:CreateWorkspaceResponse'

-> CreateWorkspaceResponse

Create a value of CreateWorkspaceResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:tags:CreateWorkspace', createWorkspaceResponse_tags - The tags of this workspace.

$sel:httpStatus:CreateWorkspaceResponse', createWorkspaceResponse_httpStatus - The response's http status code.

$sel:arn:CreateWorkspaceResponse', createWorkspaceResponse_arn - The ARN of the workspace that was just created.

$sel:status:CreateWorkspaceResponse', createWorkspaceResponse_status - The status of the workspace that was just created (usually CREATING).

$sel:workspaceId:CreateWorkspaceResponse', createWorkspaceResponse_workspaceId - The generated ID of the workspace that was just created.

Types

AlertManagerDefinitionStatusCode

newtype AlertManagerDefinitionStatusCode

State of an alert manager definition.

Constructors

AlertManagerDefinitionStatusCode'

Fields

fromAlertManagerDefinitionStatusCode :: Text

Bundled Patterns

pattern AlertManagerDefinitionStatusCode_ACTIVE :: AlertManagerDefinitionStatusCode

pattern AlertManagerDefinitionStatusCode_CREATING :: AlertManagerDefinitionStatusCode

pattern AlertManagerDefinitionStatusCode_CREATION_FAILED :: AlertManagerDefinitionStatusCode

pattern AlertManagerDefinitionStatusCode_DELETING :: AlertManagerDefinitionStatusCode

pattern AlertManagerDefinitionStatusCode_UPDATE_FAILED :: AlertManagerDefinitionStatusCode

pattern AlertManagerDefinitionStatusCode_UPDATING :: AlertManagerDefinitionStatusCode

RuleGroupsNamespaceStatusCode

newtype RuleGroupsNamespaceStatusCode

State of a namespace.

Constructors

RuleGroupsNamespaceStatusCode'

Fields

fromRuleGroupsNamespaceStatusCode :: Text

Bundled Patterns

pattern RuleGroupsNamespaceStatusCode_ACTIVE :: RuleGroupsNamespaceStatusCode

pattern RuleGroupsNamespaceStatusCode_CREATING :: RuleGroupsNamespaceStatusCode

pattern RuleGroupsNamespaceStatusCode_CREATION_FAILED :: RuleGroupsNamespaceStatusCode

pattern RuleGroupsNamespaceStatusCode_DELETING :: RuleGroupsNamespaceStatusCode

pattern RuleGroupsNamespaceStatusCode_UPDATE_FAILED :: RuleGroupsNamespaceStatusCode

pattern RuleGroupsNamespaceStatusCode_UPDATING :: RuleGroupsNamespaceStatusCode

WorkspaceStatusCode

newtype WorkspaceStatusCode

State of a workspace.

Constructors

WorkspaceStatusCode'

Fields

fromWorkspaceStatusCode :: Text

Bundled Patterns

pattern WorkspaceStatusCode_ACTIVE :: WorkspaceStatusCode

pattern WorkspaceStatusCode_CREATING :: WorkspaceStatusCode

pattern WorkspaceStatusCode_CREATION_FAILED :: WorkspaceStatusCode

pattern WorkspaceStatusCode_DELETING :: WorkspaceStatusCode

pattern WorkspaceStatusCode_UPDATING :: WorkspaceStatusCode

AlertManagerDefinitionDescription

data AlertManagerDefinitionDescription

Represents the properties of an alert manager definition.

See: newAlertManagerDefinitionDescription smart constructor.

Constructors

AlertManagerDefinitionDescription' POSIX Base64 POSIX AlertManagerDefinitionStatus

newAlertManagerDefinitionDescription

Arguments

:: UTCTime
$sel:createdAt:AlertManagerDefinitionDescription'

-> ByteString
$sel:data':AlertManagerDefinitionDescription'

-> UTCTime
$sel:modifiedAt:AlertManagerDefinitionDescription'

-> AlertManagerDefinitionStatus
$sel:status:AlertManagerDefinitionDescription'

-> AlertManagerDefinitionDescription

Create a value of AlertManagerDefinitionDescription with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:createdAt:AlertManagerDefinitionDescription', alertManagerDefinitionDescription_createdAt - The time when the alert manager definition was created.

$sel:data':AlertManagerDefinitionDescription', alertManagerDefinitionDescription_data - The alert manager definition.-- -- Note: This Lens automatically encodes and decodes Base64 data. -- The underlying isomorphism will encode to Base64 representation during -- serialisation, and decode from Base64 representation during deserialisation. -- This Lens accepts and returns only raw unencoded data.

$sel:modifiedAt:AlertManagerDefinitionDescription', alertManagerDefinitionDescription_modifiedAt - The time when the alert manager definition was modified.

$sel:status:AlertManagerDefinitionDescription', alertManagerDefinitionDescription_status - The status of alert manager definition.

AlertManagerDefinitionStatus

data AlertManagerDefinitionStatus

Represents the status of a definition.

See: newAlertManagerDefinitionStatus smart constructor.

Constructors

AlertManagerDefinitionStatus' (Maybe Text) AlertManagerDefinitionStatusCode

newAlertManagerDefinitionStatus

Arguments

:: AlertManagerDefinitionStatusCode
$sel:statusCode:AlertManagerDefinitionStatus'

-> AlertManagerDefinitionStatus

Create a value of AlertManagerDefinitionStatus with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:statusReason:AlertManagerDefinitionStatus', alertManagerDefinitionStatus_statusReason - The reason for failure if any.

$sel:statusCode:AlertManagerDefinitionStatus', alertManagerDefinitionStatus_statusCode - Status code of this definition.

RuleGroupsNamespaceDescription

data RuleGroupsNamespaceDescription

Represents a description of the rule groups namespace.

See: newRuleGroupsNamespaceDescription smart constructor.

Constructors

RuleGroupsNamespaceDescription' (Maybe (HashMap Text Text)) Text POSIX Base64 POSIX Text RuleGroupsNamespaceStatus

newRuleGroupsNamespaceDescription

Arguments

:: Text
$sel:arn:RuleGroupsNamespaceDescription'

-> UTCTime
$sel:createdAt:RuleGroupsNamespaceDescription'

-> ByteString
$sel:data':RuleGroupsNamespaceDescription'

-> UTCTime
$sel:modifiedAt:RuleGroupsNamespaceDescription'

-> Text
$sel:name:RuleGroupsNamespaceDescription'

-> RuleGroupsNamespaceStatus
$sel:status:RuleGroupsNamespaceDescription'

-> RuleGroupsNamespaceDescription

Create a value of RuleGroupsNamespaceDescription with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:tags:RuleGroupsNamespaceDescription', ruleGroupsNamespaceDescription_tags - The tags of this rule groups namespace.

$sel:arn:RuleGroupsNamespaceDescription', ruleGroupsNamespaceDescription_arn - The Amazon Resource Name (ARN) of this rule groups namespace.

$sel:createdAt:RuleGroupsNamespaceDescription', ruleGroupsNamespaceDescription_createdAt - The time when the rule groups namespace was created.

$sel:data':RuleGroupsNamespaceDescription', ruleGroupsNamespaceDescription_data - The rule groups namespace data.-- -- Note: This Lens automatically encodes and decodes Base64 data. -- The underlying isomorphism will encode to Base64 representation during -- serialisation, and decode from Base64 representation during deserialisation. -- This Lens accepts and returns only raw unencoded data.

$sel:modifiedAt:RuleGroupsNamespaceDescription', ruleGroupsNamespaceDescription_modifiedAt - The time when the rule groups namespace was modified.

$sel:name:RuleGroupsNamespaceDescription', ruleGroupsNamespaceDescription_name - The rule groups namespace name.

$sel:status:RuleGroupsNamespaceDescription', ruleGroupsNamespaceDescription_status - The status of rule groups namespace.

RuleGroupsNamespaceStatus

data RuleGroupsNamespaceStatus

Represents the status of a namespace.

See: newRuleGroupsNamespaceStatus smart constructor.

Constructors

RuleGroupsNamespaceStatus' (Maybe Text) RuleGroupsNamespaceStatusCode

newRuleGroupsNamespaceStatus

Arguments

:: RuleGroupsNamespaceStatusCode
$sel:statusCode:RuleGroupsNamespaceStatus'

-> RuleGroupsNamespaceStatus

Create a value of RuleGroupsNamespaceStatus with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:statusReason:RuleGroupsNamespaceStatus', ruleGroupsNamespaceStatus_statusReason - The reason for failure if any.

$sel:statusCode:RuleGroupsNamespaceStatus', ruleGroupsNamespaceStatus_statusCode - Status code of this namespace.

RuleGroupsNamespaceSummary

data RuleGroupsNamespaceSummary

Represents a summary of the rule groups namespace.

See: newRuleGroupsNamespaceSummary smart constructor.

Constructors

RuleGroupsNamespaceSummary' (Maybe (HashMap Text Text)) Text POSIX POSIX Text RuleGroupsNamespaceStatus

newRuleGroupsNamespaceSummary

Arguments

:: Text
$sel:arn:RuleGroupsNamespaceSummary'

-> UTCTime
$sel:createdAt:RuleGroupsNamespaceSummary'

-> UTCTime
$sel:modifiedAt:RuleGroupsNamespaceSummary'

-> Text
$sel:name:RuleGroupsNamespaceSummary'

-> RuleGroupsNamespaceStatus
$sel:status:RuleGroupsNamespaceSummary'

-> RuleGroupsNamespaceSummary

Create a value of RuleGroupsNamespaceSummary with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:tags:RuleGroupsNamespaceSummary', ruleGroupsNamespaceSummary_tags - The tags of this rule groups namespace.

$sel:arn:RuleGroupsNamespaceSummary', ruleGroupsNamespaceSummary_arn - The Amazon Resource Name (ARN) of this rule groups namespace.

$sel:createdAt:RuleGroupsNamespaceSummary', ruleGroupsNamespaceSummary_createdAt - The time when the rule groups namespace was created.

$sel:modifiedAt:RuleGroupsNamespaceSummary', ruleGroupsNamespaceSummary_modifiedAt - The time when the rule groups namespace was modified.

$sel:name:RuleGroupsNamespaceSummary', ruleGroupsNamespaceSummary_name - The rule groups namespace name.

$sel:status:RuleGroupsNamespaceSummary', ruleGroupsNamespaceSummary_status - The status of rule groups namespace.

WorkspaceDescription

data WorkspaceDescription

Represents the properties of a workspace.

See: newWorkspaceDescription smart constructor.

Constructors

WorkspaceDescription' (Maybe Text) (Maybe Text) (Maybe (HashMap Text Text)) Text POSIX WorkspaceStatus Text

newWorkspaceDescription

Arguments

:: Text
$sel:arn:WorkspaceDescription'

-> UTCTime
$sel:createdAt:WorkspaceDescription'

-> WorkspaceStatus
$sel:status:WorkspaceDescription'

-> Text
$sel:workspaceId:WorkspaceDescription'

-> WorkspaceDescription

Create a value of WorkspaceDescription with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:alias:WorkspaceDescription', workspaceDescription_alias - Alias of this workspace.

$sel:prometheusEndpoint:WorkspaceDescription', workspaceDescription_prometheusEndpoint - Prometheus endpoint URI.

$sel:tags:WorkspaceDescription', workspaceDescription_tags - The tags of this workspace.

$sel:arn:WorkspaceDescription', workspaceDescription_arn - The Amazon Resource Name (ARN) of this workspace.

$sel:createdAt:WorkspaceDescription', workspaceDescription_createdAt - The time when the workspace was created.

$sel:status:WorkspaceDescription', workspaceDescription_status - The status of this workspace.

$sel:workspaceId:WorkspaceDescription', workspaceDescription_workspaceId - Unique string identifying this workspace.

WorkspaceStatus

data WorkspaceStatus

Represents the status of a workspace.

See: newWorkspaceStatus smart constructor.

Constructors

WorkspaceStatus' WorkspaceStatusCode

newWorkspaceStatus

Arguments

:: WorkspaceStatusCode
$sel:statusCode:WorkspaceStatus'

-> WorkspaceStatus

Create a value of WorkspaceStatus with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:statusCode:WorkspaceStatus', workspaceStatus_statusCode - Status code of this workspace.

WorkspaceSummary

data WorkspaceSummary

Represents a summary of the properties of a workspace.

See: newWorkspaceSummary smart constructor.

Constructors

WorkspaceSummary' (Maybe Text) (Maybe (HashMap Text Text)) Text POSIX WorkspaceStatus Text

newWorkspaceSummary

Arguments

:: Text
$sel:arn:WorkspaceSummary'

-> UTCTime
$sel:createdAt:WorkspaceSummary'

-> WorkspaceStatus
$sel:status:WorkspaceSummary'

-> Text
$sel:workspaceId:WorkspaceSummary'

-> WorkspaceSummary

Create a value of WorkspaceSummary with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:alias:WorkspaceSummary', workspaceSummary_alias - Alias of this workspace.

$sel:tags:WorkspaceSummary', workspaceSummary_tags - The tags of this workspace.

$sel:arn:WorkspaceSummary', workspaceSummary_arn - The AmazonResourceName of this workspace.

$sel:createdAt:WorkspaceSummary', workspaceSummary_createdAt - The time when the workspace was created.

$sel:status:WorkspaceSummary', workspaceSummary_status - The status of this workspace.

$sel:workspaceId:WorkspaceSummary', workspaceSummary_workspaceId - Unique string identifying this workspace.