Amazonka.AMP.Lens

Description

Operations

ListWorkspaces

listWorkspaces_alias :: Lens' ListWorkspaces (Maybe Text)

Optional filter for workspace alias. Only the workspaces with aliases that begin with this value will be returned.

listWorkspaces_nextToken :: Lens' ListWorkspaces (Maybe Text)

Pagination token to request the next page in a paginated list. This token is obtained from the output of the previous ListWorkspaces request.

listWorkspaces_maxResults :: Lens' ListWorkspaces (Maybe Natural)

Maximum results to return in response (default=100, maximum=1000).

listWorkspacesResponse_nextToken :: Lens' ListWorkspacesResponse (Maybe Text)

Pagination token to use when requesting the next page in this list.

listWorkspacesResponse_httpStatus :: Lens' ListWorkspacesResponse Int

The response's http status code.

listWorkspacesResponse_workspaces :: Lens' ListWorkspacesResponse [WorkspaceSummary]

The list of existing workspaces, including those undergoing creation or deletion.

CreateAlertManagerDefinition

createAlertManagerDefinition_clientToken :: Lens' CreateAlertManagerDefinition (Maybe Text)

Optional, unique, case-sensitive, user-provided identifier to ensure the idempotency of the request.

createAlertManagerDefinition_data :: Lens' CreateAlertManagerDefinition ByteString

The alert manager definition data.-- -- Note: This Lens automatically encodes and decodes Base64 data. -- The underlying isomorphism will encode to Base64 representation during -- serialisation, and decode from Base64 representation during deserialisation. -- This Lens accepts and returns only raw unencoded data.

createAlertManagerDefinition_workspaceId :: Lens' CreateAlertManagerDefinition Text

The ID of the workspace in which to create the alert manager definition.

createAlertManagerDefinitionResponse_httpStatus :: Lens' CreateAlertManagerDefinitionResponse Int

The response's http status code.

createAlertManagerDefinitionResponse_status :: Lens' CreateAlertManagerDefinitionResponse AlertManagerDefinitionStatus

The status of alert manager definition.

DeleteWorkspace

deleteWorkspace_clientToken :: Lens' DeleteWorkspace (Maybe Text)

Optional, unique, case-sensitive, user-provided identifier to ensure the idempotency of the request.

deleteWorkspace_workspaceId :: Lens' DeleteWorkspace Text

The ID of the workspace to delete.

ListTagsForResource

listTagsForResource_resourceArn :: Lens' ListTagsForResource Text

The ARN of the resource.

listTagsForResourceResponse_tags :: Lens' ListTagsForResourceResponse (Maybe (HashMap Text Text))

Undocumented member.

listTagsForResourceResponse_httpStatus :: Lens' ListTagsForResourceResponse Int

The response's http status code.

CreateRuleGroupsNamespace

createRuleGroupsNamespace_clientToken :: Lens' CreateRuleGroupsNamespace (Maybe Text)

Optional, unique, case-sensitive, user-provided identifier to ensure the idempotency of the request.

createRuleGroupsNamespace_tags :: Lens' CreateRuleGroupsNamespace (Maybe (HashMap Text Text))

Optional, user-provided tags for this rule groups namespace.

createRuleGroupsNamespace_data :: Lens' CreateRuleGroupsNamespace ByteString

The namespace data that define the rule groups.-- -- Note: This Lens automatically encodes and decodes Base64 data. -- The underlying isomorphism will encode to Base64 representation during -- serialisation, and decode from Base64 representation during deserialisation. -- This Lens accepts and returns only raw unencoded data.

createRuleGroupsNamespace_name :: Lens' CreateRuleGroupsNamespace Text

The rule groups namespace name.

createRuleGroupsNamespace_workspaceId :: Lens' CreateRuleGroupsNamespace Text

The ID of the workspace in which to create the rule group namespace.

createRuleGroupsNamespaceResponse_tags :: Lens' CreateRuleGroupsNamespaceResponse (Maybe (HashMap Text Text))

The tags of this rule groups namespace.

createRuleGroupsNamespaceResponse_httpStatus :: Lens' CreateRuleGroupsNamespaceResponse Int

The response's http status code.

createRuleGroupsNamespaceResponse_arn :: Lens' CreateRuleGroupsNamespaceResponse Text

The Amazon Resource Name (ARN) of this rule groups namespace.

createRuleGroupsNamespaceResponse_name :: Lens' CreateRuleGroupsNamespaceResponse Text

The rule groups namespace name.

createRuleGroupsNamespaceResponse_status :: Lens' CreateRuleGroupsNamespaceResponse RuleGroupsNamespaceStatus

The status of rule groups namespace.

DescribeAlertManagerDefinition

describeAlertManagerDefinition_workspaceId :: Lens' DescribeAlertManagerDefinition Text

The ID of the workspace to describe.

describeAlertManagerDefinitionResponse_httpStatus :: Lens' DescribeAlertManagerDefinitionResponse Int

The response's http status code.

describeAlertManagerDefinitionResponse_alertManagerDefinition :: Lens' DescribeAlertManagerDefinitionResponse AlertManagerDefinitionDescription

The properties of the selected workspace's alert manager definition.

DescribeWorkspace

describeWorkspace_workspaceId :: Lens' DescribeWorkspace Text

The ID of the workspace to describe.

describeWorkspaceResponse_httpStatus :: Lens' DescribeWorkspaceResponse Int

The response's http status code.

describeWorkspaceResponse_workspace :: Lens' DescribeWorkspaceResponse WorkspaceDescription

The properties of the selected workspace.

PutAlertManagerDefinition

putAlertManagerDefinition_clientToken :: Lens' PutAlertManagerDefinition (Maybe Text)

Optional, unique, case-sensitive, user-provided identifier to ensure the idempotency of the request.

putAlertManagerDefinition_data :: Lens' PutAlertManagerDefinition ByteString

The alert manager definition data.-- -- Note: This Lens automatically encodes and decodes Base64 data. -- The underlying isomorphism will encode to Base64 representation during -- serialisation, and decode from Base64 representation during deserialisation. -- This Lens accepts and returns only raw unencoded data.

putAlertManagerDefinition_workspaceId :: Lens' PutAlertManagerDefinition Text

The ID of the workspace in which to update the alert manager definition.

putAlertManagerDefinitionResponse_httpStatus :: Lens' PutAlertManagerDefinitionResponse Int

The response's http status code.

putAlertManagerDefinitionResponse_status :: Lens' PutAlertManagerDefinitionResponse AlertManagerDefinitionStatus

The status of alert manager definition.

DeleteAlertManagerDefinition

deleteAlertManagerDefinition_clientToken :: Lens' DeleteAlertManagerDefinition (Maybe Text)

Optional, unique, case-sensitive, user-provided identifier to ensure the idempotency of the request.

deleteAlertManagerDefinition_workspaceId :: Lens' DeleteAlertManagerDefinition Text

The ID of the workspace in which to delete the alert manager definition.

DescribeRuleGroupsNamespace

describeRuleGroupsNamespace_name :: Lens' DescribeRuleGroupsNamespace Text

The rule groups namespace.

describeRuleGroupsNamespace_workspaceId :: Lens' DescribeRuleGroupsNamespace Text

The ID of the workspace to describe.

describeRuleGroupsNamespaceResponse_httpStatus :: Lens' DescribeRuleGroupsNamespaceResponse Int

The response's http status code.

describeRuleGroupsNamespaceResponse_ruleGroupsNamespace :: Lens' DescribeRuleGroupsNamespaceResponse RuleGroupsNamespaceDescription

The selected rule groups namespace.

UpdateWorkspaceAlias

updateWorkspaceAlias_clientToken :: Lens' UpdateWorkspaceAlias (Maybe Text)

Optional, unique, case-sensitive, user-provided identifier to ensure the idempotency of the request.

updateWorkspaceAlias_alias :: Lens' UpdateWorkspaceAlias (Maybe Text)

The new alias of the workspace.

updateWorkspaceAlias_workspaceId :: Lens' UpdateWorkspaceAlias Text

The ID of the workspace being updated.

DeleteRuleGroupsNamespace

deleteRuleGroupsNamespace_clientToken :: Lens' DeleteRuleGroupsNamespace (Maybe Text)

Optional, unique, case-sensitive, user-provided identifier to ensure the idempotency of the request.

deleteRuleGroupsNamespace_name :: Lens' DeleteRuleGroupsNamespace Text

The rule groups namespace name.

deleteRuleGroupsNamespace_workspaceId :: Lens' DeleteRuleGroupsNamespace Text

The ID of the workspace to delete rule group definition.

PutRuleGroupsNamespace

putRuleGroupsNamespace_clientToken :: Lens' PutRuleGroupsNamespace (Maybe Text)

Optional, unique, case-sensitive, user-provided identifier to ensure the idempotency of the request.

putRuleGroupsNamespace_data :: Lens' PutRuleGroupsNamespace ByteString

The namespace data that define the rule groups.-- -- Note: This Lens automatically encodes and decodes Base64 data. -- The underlying isomorphism will encode to Base64 representation during -- serialisation, and decode from Base64 representation during deserialisation. -- This Lens accepts and returns only raw unencoded data.

putRuleGroupsNamespace_name :: Lens' PutRuleGroupsNamespace Text

The rule groups namespace name.

putRuleGroupsNamespace_workspaceId :: Lens' PutRuleGroupsNamespace Text

The ID of the workspace in which to update the rule group namespace.

putRuleGroupsNamespaceResponse_tags :: Lens' PutRuleGroupsNamespaceResponse (Maybe (HashMap Text Text))

The tags of this rule groups namespace.

putRuleGroupsNamespaceResponse_httpStatus :: Lens' PutRuleGroupsNamespaceResponse Int

The response's http status code.

putRuleGroupsNamespaceResponse_arn :: Lens' PutRuleGroupsNamespaceResponse Text

The Amazon Resource Name (ARN) of this rule groups namespace.

putRuleGroupsNamespaceResponse_name :: Lens' PutRuleGroupsNamespaceResponse Text

The rule groups namespace name.

putRuleGroupsNamespaceResponse_status :: Lens' PutRuleGroupsNamespaceResponse RuleGroupsNamespaceStatus

The status of rule groups namespace.

ListRuleGroupsNamespaces

listRuleGroupsNamespaces_nextToken :: Lens' ListRuleGroupsNamespaces (Maybe Text)

Pagination token to request the next page in a paginated list. This token is obtained from the output of the previous ListRuleGroupsNamespaces request.

listRuleGroupsNamespaces_name :: Lens' ListRuleGroupsNamespaces (Maybe Text)

Optional filter for rule groups namespace name. Only the rule groups namespace that begin with this value will be returned.

listRuleGroupsNamespaces_maxResults :: Lens' ListRuleGroupsNamespaces (Maybe Natural)

Maximum results to return in response (default=100, maximum=1000).

listRuleGroupsNamespaces_workspaceId :: Lens' ListRuleGroupsNamespaces Text

The ID of the workspace.

listRuleGroupsNamespacesResponse_nextToken :: Lens' ListRuleGroupsNamespacesResponse (Maybe Text)

Pagination token to use when requesting the next page in this list.

listRuleGroupsNamespacesResponse_httpStatus :: Lens' ListRuleGroupsNamespacesResponse Int

The response's http status code.

listRuleGroupsNamespacesResponse_ruleGroupsNamespaces :: Lens' ListRuleGroupsNamespacesResponse [RuleGroupsNamespaceSummary]

The list of the selected rule groups namespaces.

TagResource

tagResource_resourceArn :: Lens' TagResource Text

The ARN of the resource.

tagResource_tags :: Lens' TagResource (HashMap Text Text)

Undocumented member.

tagResourceResponse_httpStatus :: Lens' TagResourceResponse Int

The response's http status code.

UntagResource

untagResource_resourceArn :: Lens' UntagResource Text

The ARN of the resource.

untagResource_tagKeys :: Lens' UntagResource [Text]

One or more tag keys

untagResourceResponse_httpStatus :: Lens' UntagResourceResponse Int

The response's http status code.

CreateWorkspace

createWorkspace_clientToken :: Lens' CreateWorkspace (Maybe Text)

Optional, unique, case-sensitive, user-provided identifier to ensure the idempotency of the request.

createWorkspace_alias :: Lens' CreateWorkspace (Maybe Text)

An optional user-assigned alias for this workspace. This alias is for user reference and does not need to be unique.

createWorkspace_tags :: Lens' CreateWorkspace (Maybe (HashMap Text Text))

Optional, user-provided tags for this workspace.

createWorkspaceResponse_tags :: Lens' CreateWorkspaceResponse (Maybe (HashMap Text Text))

The tags of this workspace.

createWorkspaceResponse_httpStatus :: Lens' CreateWorkspaceResponse Int

The response's http status code.

createWorkspaceResponse_arn :: Lens' CreateWorkspaceResponse Text

The ARN of the workspace that was just created.

createWorkspaceResponse_status :: Lens' CreateWorkspaceResponse WorkspaceStatus

The status of the workspace that was just created (usually CREATING).

createWorkspaceResponse_workspaceId :: Lens' CreateWorkspaceResponse Text

The generated ID of the workspace that was just created.

Types

AlertManagerDefinitionDescription

alertManagerDefinitionDescription_createdAt :: Lens' AlertManagerDefinitionDescription UTCTime

The time when the alert manager definition was created.

alertManagerDefinitionDescription_data :: Lens' AlertManagerDefinitionDescription ByteString

The alert manager definition.-- -- Note: This Lens automatically encodes and decodes Base64 data. -- The underlying isomorphism will encode to Base64 representation during -- serialisation, and decode from Base64 representation during deserialisation. -- This Lens accepts and returns only raw unencoded data.

alertManagerDefinitionDescription_modifiedAt :: Lens' AlertManagerDefinitionDescription UTCTime

The time when the alert manager definition was modified.

alertManagerDefinitionDescription_status :: Lens' AlertManagerDefinitionDescription AlertManagerDefinitionStatus

The status of alert manager definition.

AlertManagerDefinitionStatus

alertManagerDefinitionStatus_statusReason :: Lens' AlertManagerDefinitionStatus (Maybe Text)

The reason for failure if any.

alertManagerDefinitionStatus_statusCode :: Lens' AlertManagerDefinitionStatus AlertManagerDefinitionStatusCode

Status code of this definition.

RuleGroupsNamespaceDescription

ruleGroupsNamespaceDescription_tags :: Lens' RuleGroupsNamespaceDescription (Maybe (HashMap Text Text))

The tags of this rule groups namespace.

ruleGroupsNamespaceDescription_arn :: Lens' RuleGroupsNamespaceDescription Text

The Amazon Resource Name (ARN) of this rule groups namespace.

ruleGroupsNamespaceDescription_createdAt :: Lens' RuleGroupsNamespaceDescription UTCTime

The time when the rule groups namespace was created.

ruleGroupsNamespaceDescription_data :: Lens' RuleGroupsNamespaceDescription ByteString

The rule groups namespace data.-- -- Note: This Lens automatically encodes and decodes Base64 data. -- The underlying isomorphism will encode to Base64 representation during -- serialisation, and decode from Base64 representation during deserialisation. -- This Lens accepts and returns only raw unencoded data.

ruleGroupsNamespaceDescription_modifiedAt :: Lens' RuleGroupsNamespaceDescription UTCTime

The time when the rule groups namespace was modified.

ruleGroupsNamespaceDescription_name :: Lens' RuleGroupsNamespaceDescription Text

The rule groups namespace name.

ruleGroupsNamespaceDescription_status :: Lens' RuleGroupsNamespaceDescription RuleGroupsNamespaceStatus

The status of rule groups namespace.

RuleGroupsNamespaceStatus

ruleGroupsNamespaceStatus_statusReason :: Lens' RuleGroupsNamespaceStatus (Maybe Text)

The reason for failure if any.

ruleGroupsNamespaceStatus_statusCode :: Lens' RuleGroupsNamespaceStatus RuleGroupsNamespaceStatusCode

Status code of this namespace.

RuleGroupsNamespaceSummary

ruleGroupsNamespaceSummary_tags :: Lens' RuleGroupsNamespaceSummary (Maybe (HashMap Text Text))

The tags of this rule groups namespace.

ruleGroupsNamespaceSummary_arn :: Lens' RuleGroupsNamespaceSummary Text

The Amazon Resource Name (ARN) of this rule groups namespace.

ruleGroupsNamespaceSummary_createdAt :: Lens' RuleGroupsNamespaceSummary UTCTime

The time when the rule groups namespace was created.

ruleGroupsNamespaceSummary_modifiedAt :: Lens' RuleGroupsNamespaceSummary UTCTime

The time when the rule groups namespace was modified.

ruleGroupsNamespaceSummary_name :: Lens' RuleGroupsNamespaceSummary Text

The rule groups namespace name.

ruleGroupsNamespaceSummary_status :: Lens' RuleGroupsNamespaceSummary RuleGroupsNamespaceStatus

The status of rule groups namespace.

WorkspaceDescription

workspaceDescription_alias :: Lens' WorkspaceDescription (Maybe Text)

Alias of this workspace.

workspaceDescription_prometheusEndpoint :: Lens' WorkspaceDescription (Maybe Text)

Prometheus endpoint URI.

workspaceDescription_tags :: Lens' WorkspaceDescription (Maybe (HashMap Text Text))

The tags of this workspace.

workspaceDescription_arn :: Lens' WorkspaceDescription Text

The Amazon Resource Name (ARN) of this workspace.

workspaceDescription_createdAt :: Lens' WorkspaceDescription UTCTime

The time when the workspace was created.

workspaceDescription_status :: Lens' WorkspaceDescription WorkspaceStatus

The status of this workspace.

workspaceDescription_workspaceId :: Lens' WorkspaceDescription Text

Unique string identifying this workspace.

WorkspaceStatus

workspaceStatus_statusCode :: Lens' WorkspaceStatus WorkspaceStatusCode

Status code of this workspace.

WorkspaceSummary

workspaceSummary_alias :: Lens' WorkspaceSummary (Maybe Text)

Alias of this workspace.

workspaceSummary_tags :: Lens' WorkspaceSummary (Maybe (HashMap Text Text))

The tags of this workspace.

workspaceSummary_arn :: Lens' WorkspaceSummary Text

The AmazonResourceName of this workspace.

workspaceSummary_createdAt :: Lens' WorkspaceSummary UTCTime

The time when the workspace was created.

workspaceSummary_status :: Lens' WorkspaceSummary WorkspaceStatus

The status of this workspace.

workspaceSummary_workspaceId :: Lens' WorkspaceSummary Text

Unique string identifying this workspace.