Amazonka.AMP.PutAlertManagerDefinition

Description

Update an alert manager definition.

Creating a Request

data PutAlertManagerDefinition

Represents the input of a PutAlertManagerDefinition operation.

See: newPutAlertManagerDefinition smart constructor.

Constructors

PutAlertManagerDefinition'

Fields

clientToken :: Maybe Text

Optional, unique, case-sensitive, user-provided identifier to ensure the idempotency of the request.

data' :: Base64

The alert manager definition data.

workspaceId :: Text

The ID of the workspace in which to update the alert manager definition.

newPutAlertManagerDefinition

Arguments

:: ByteString
$sel:data':PutAlertManagerDefinition'

-> Text
$sel:workspaceId:PutAlertManagerDefinition'

-> PutAlertManagerDefinition

Create a value of PutAlertManagerDefinition with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:clientToken:PutAlertManagerDefinition', putAlertManagerDefinition_clientToken - Optional, unique, case-sensitive, user-provided identifier to ensure the idempotency of the request.

$sel:data':PutAlertManagerDefinition', putAlertManagerDefinition_data - The alert manager definition data.-- -- Note: This Lens automatically encodes and decodes Base64 data. -- The underlying isomorphism will encode to Base64 representation during -- serialisation, and decode from Base64 representation during deserialisation. -- This Lens accepts and returns only raw unencoded data.

$sel:workspaceId:PutAlertManagerDefinition', putAlertManagerDefinition_workspaceId - The ID of the workspace in which to update the alert manager definition.

Request Lenses

putAlertManagerDefinition_clientToken :: Lens' PutAlertManagerDefinition (Maybe Text)

Optional, unique, case-sensitive, user-provided identifier to ensure the idempotency of the request.

putAlertManagerDefinition_data :: Lens' PutAlertManagerDefinition ByteString

The alert manager definition data.-- -- Note: This Lens automatically encodes and decodes Base64 data. -- The underlying isomorphism will encode to Base64 representation during -- serialisation, and decode from Base64 representation during deserialisation. -- This Lens accepts and returns only raw unencoded data.

putAlertManagerDefinition_workspaceId :: Lens' PutAlertManagerDefinition Text

The ID of the workspace in which to update the alert manager definition.

Destructuring the Response

data PutAlertManagerDefinitionResponse

Represents the output of a PutAlertManagerDefinition operation.

See: newPutAlertManagerDefinitionResponse smart constructor.

Constructors

PutAlertManagerDefinitionResponse'

Fields

httpStatus :: Int

The response's http status code.

status :: AlertManagerDefinitionStatus

The status of alert manager definition.

newPutAlertManagerDefinitionResponse

Arguments

:: Int
$sel:httpStatus:PutAlertManagerDefinitionResponse'

-> AlertManagerDefinitionStatus
$sel:status:PutAlertManagerDefinitionResponse'

-> PutAlertManagerDefinitionResponse

Create a value of PutAlertManagerDefinitionResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:httpStatus:PutAlertManagerDefinitionResponse', putAlertManagerDefinitionResponse_httpStatus - The response's http status code.

$sel:status:PutAlertManagerDefinitionResponse', putAlertManagerDefinitionResponse_status - The status of alert manager definition.

Response Lenses

putAlertManagerDefinitionResponse_httpStatus :: Lens' PutAlertManagerDefinitionResponse Int

The response's http status code.

putAlertManagerDefinitionResponse_status :: Lens' PutAlertManagerDefinitionResponse AlertManagerDefinitionStatus

The status of alert manager definition.