Amazonka.AMP.Types.WorkspaceSummary

Description

Documentation

data WorkspaceSummary

Represents a summary of the properties of a workspace.

See: newWorkspaceSummary smart constructor.

Constructors

WorkspaceSummary'

Fields

alias :: Maybe Text

Alias of this workspace.

tags :: Maybe (HashMap Text Text)

The tags of this workspace.

arn :: Text

The AmazonResourceName of this workspace.

createdAt :: POSIX

The time when the workspace was created.

status :: WorkspaceStatus

The status of this workspace.

workspaceId :: Text

Unique string identifying this workspace.

newWorkspaceSummary

Arguments

:: Text
$sel:arn:WorkspaceSummary'

-> UTCTime
$sel:createdAt:WorkspaceSummary'

-> WorkspaceStatus
$sel:status:WorkspaceSummary'

-> Text
$sel:workspaceId:WorkspaceSummary'

-> WorkspaceSummary

Create a value of WorkspaceSummary with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:alias:WorkspaceSummary', workspaceSummary_alias - Alias of this workspace.

$sel:tags:WorkspaceSummary', workspaceSummary_tags - The tags of this workspace.

$sel:arn:WorkspaceSummary', workspaceSummary_arn - The AmazonResourceName of this workspace.

$sel:createdAt:WorkspaceSummary', workspaceSummary_createdAt - The time when the workspace was created.

$sel:status:WorkspaceSummary', workspaceSummary_status - The status of this workspace.

$sel:workspaceId:WorkspaceSummary', workspaceSummary_workspaceId - Unique string identifying this workspace.

workspaceSummary_alias :: Lens' WorkspaceSummary (Maybe Text)

Alias of this workspace.

workspaceSummary_tags :: Lens' WorkspaceSummary (Maybe (HashMap Text Text))

The tags of this workspace.

workspaceSummary_arn :: Lens' WorkspaceSummary Text

The AmazonResourceName of this workspace.

workspaceSummary_createdAt :: Lens' WorkspaceSummary UTCTime

The time when the workspace was created.

workspaceSummary_status :: Lens' WorkspaceSummary WorkspaceStatus

The status of this workspace.

workspaceSummary_workspaceId :: Lens' WorkspaceSummary Text

Unique string identifying this workspace.