Amazonka.APIGateway.CreateBasePathMapping

Description

Creates a new BasePathMapping resource.

Creating a Request

data CreateBasePathMapping

Requests API Gateway to create a new BasePathMapping resource.

See: newCreateBasePathMapping smart constructor.

Constructors

CreateBasePathMapping'

Fields

stage :: Maybe Text

The name of the API's stage that you want to use for this mapping. Specify '(none)' if you want callers to explicitly specify the stage name after any base path name.

basePath :: Maybe Text

The base path name that callers of the API must provide as part of the URL after the domain name. This value must be unique for all of the mappings across a single API. Specify '(none)' if you do not want callers to specify a base path name after the domain name.

domainName :: Text

Required

The domain name of the BasePathMapping resource to create.

restApiId :: Text

Required

The string identifier of the associated RestApi.

newCreateBasePathMapping

Arguments

:: Text
$sel:domainName:CreateBasePathMapping'

-> Text
$sel:restApiId:CreateBasePathMapping'

-> CreateBasePathMapping

Create a value of CreateBasePathMapping with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:stage:CreateBasePathMapping', createBasePathMapping_stage - The name of the API's stage that you want to use for this mapping. Specify '(none)' if you want callers to explicitly specify the stage name after any base path name.

$sel:basePath:CreateBasePathMapping', createBasePathMapping_basePath - The base path name that callers of the API must provide as part of the URL after the domain name. This value must be unique for all of the mappings across a single API. Specify '(none)' if you do not want callers to specify a base path name after the domain name.

$sel:domainName:CreateBasePathMapping', createBasePathMapping_domainName - [Required] The domain name of the BasePathMapping resource to create.

$sel:restApiId:CreateBasePathMapping', createBasePathMapping_restApiId - [Required] The string identifier of the associated RestApi.

Request Lenses

createBasePathMapping_stage :: Lens' CreateBasePathMapping (Maybe Text)

The name of the API's stage that you want to use for this mapping. Specify '(none)' if you want callers to explicitly specify the stage name after any base path name.

createBasePathMapping_basePath :: Lens' CreateBasePathMapping (Maybe Text)

The base path name that callers of the API must provide as part of the URL after the domain name. This value must be unique for all of the mappings across a single API. Specify '(none)' if you do not want callers to specify a base path name after the domain name.

createBasePathMapping_domainName :: Lens' CreateBasePathMapping Text

Required

The domain name of the BasePathMapping resource to create.

createBasePathMapping_restApiId :: Lens' CreateBasePathMapping Text

Required

The string identifier of the associated RestApi.

Destructuring the Response

data BasePathMapping

Represents the base path that callers of the API must provide as part of the URL after the domain name.

A custom domain name plus a BasePathMapping specification identifies a deployed RestApi in a given stage of the owner Account.

Use Custom Domain Names

See: newBasePathMapping smart constructor.

Constructors

BasePathMapping'

Fields

stage :: Maybe Text

The name of the associated stage.

basePath :: Maybe Text

The base path name that callers of the API must provide as part of the URL after the domain name.

restApiId :: Maybe Text

The string identifier of the associated RestApi.

newBasePathMapping :: BasePathMapping

Create a value of BasePathMapping with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:stage:BasePathMapping', basePathMapping_stage - The name of the associated stage.

$sel:basePath:BasePathMapping', basePathMapping_basePath - The base path name that callers of the API must provide as part of the URL after the domain name.

$sel:restApiId:BasePathMapping', basePathMapping_restApiId - The string identifier of the associated RestApi.

Response Lenses

basePathMapping_stage :: Lens' BasePathMapping (Maybe Text)

The name of the associated stage.

basePathMapping_basePath :: Lens' BasePathMapping (Maybe Text)

The base path name that callers of the API must provide as part of the URL after the domain name.

basePathMapping_restApiId :: Lens' BasePathMapping (Maybe Text)

The string identifier of the associated RestApi.