Amazonka.APIGateway.GetBasePathMappings

Description

Represents a collection of BasePathMapping resources.

This operation returns paginated results.

Creating a Request

data GetBasePathMappings

A request to get information about a collection of BasePathMapping resources.

See: newGetBasePathMappings smart constructor.

Constructors

GetBasePathMappings'

Fields

limit :: Maybe Int

The maximum number of returned results per page. The default value is 25 and the maximum value is 500.

position :: Maybe Text

The current pagination position in the paged result set.

domainName :: Text

Required

The domain name of a BasePathMapping resource.

newGetBasePathMappings

Arguments

:: Text
$sel:domainName:GetBasePathMappings'

-> GetBasePathMappings

Create a value of GetBasePathMappings with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:limit:GetBasePathMappings', getBasePathMappings_limit - The maximum number of returned results per page. The default value is 25 and the maximum value is 500.

$sel:position:GetBasePathMappings', getBasePathMappings_position - The current pagination position in the paged result set.

$sel:domainName:GetBasePathMappings', getBasePathMappings_domainName - [Required] The domain name of a BasePathMapping resource.

Request Lenses

getBasePathMappings_limit :: Lens' GetBasePathMappings (Maybe Int)

The maximum number of returned results per page. The default value is 25 and the maximum value is 500.

getBasePathMappings_position :: Lens' GetBasePathMappings (Maybe Text)

The current pagination position in the paged result set.

getBasePathMappings_domainName :: Lens' GetBasePathMappings Text

Required

The domain name of a BasePathMapping resource.

Destructuring the Response

data GetBasePathMappingsResponse

Represents a collection of BasePathMapping resources.

Use Custom Domain Names

See: newGetBasePathMappingsResponse smart constructor.

Constructors

GetBasePathMappingsResponse'

Fields

items :: Maybe [BasePathMapping]

The current page of elements from this collection.

position :: Maybe Text

httpStatus :: Int

The response's http status code.

newGetBasePathMappingsResponse

Arguments

:: Int
$sel:httpStatus:GetBasePathMappingsResponse'

-> GetBasePathMappingsResponse

Create a value of GetBasePathMappingsResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:items:GetBasePathMappingsResponse', getBasePathMappingsResponse_items - The current page of elements from this collection.

$sel:position:GetBasePathMappings', getBasePathMappingsResponse_position - Undocumented member.

$sel:httpStatus:GetBasePathMappingsResponse', getBasePathMappingsResponse_httpStatus - The response's http status code.

Response Lenses

getBasePathMappingsResponse_items :: Lens' GetBasePathMappingsResponse (Maybe [BasePathMapping])

The current page of elements from this collection.

getBasePathMappingsResponse_position :: Lens' GetBasePathMappingsResponse (Maybe Text)

Undocumented member.

getBasePathMappingsResponse_httpStatus :: Lens' GetBasePathMappingsResponse Int

The response's http status code.