Amazonka.APIGateway.GetClientCertificates

Description

Gets a collection of ClientCertificate resources.

This operation returns paginated results.

Creating a Request

data GetClientCertificates

A request to get information about a collection of ClientCertificate resources.

See: newGetClientCertificates smart constructor.

Constructors

GetClientCertificates'

Fields

limit :: Maybe Int

The maximum number of returned results per page. The default value is 25 and the maximum value is 500.

position :: Maybe Text

The current pagination position in the paged result set.

newGetClientCertificates :: GetClientCertificates

Create a value of GetClientCertificates with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:limit:GetClientCertificates', getClientCertificates_limit - The maximum number of returned results per page. The default value is 25 and the maximum value is 500.

$sel:position:GetClientCertificates', getClientCertificates_position - The current pagination position in the paged result set.

Request Lenses

getClientCertificates_limit :: Lens' GetClientCertificates (Maybe Int)

The maximum number of returned results per page. The default value is 25 and the maximum value is 500.

getClientCertificates_position :: Lens' GetClientCertificates (Maybe Text)

The current pagination position in the paged result set.

Destructuring the Response

data GetClientCertificatesResponse

Represents a collection of ClientCertificate resources.

Use Client-Side Certificate

See: newGetClientCertificatesResponse smart constructor.

Constructors

GetClientCertificatesResponse'

Fields

items :: Maybe [ClientCertificate]

The current page of elements from this collection.

position :: Maybe Text

httpStatus :: Int

The response's http status code.

newGetClientCertificatesResponse

Arguments

:: Int
$sel:httpStatus:GetClientCertificatesResponse'

-> GetClientCertificatesResponse

Create a value of GetClientCertificatesResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:items:GetClientCertificatesResponse', getClientCertificatesResponse_items - The current page of elements from this collection.

$sel:position:GetClientCertificates', getClientCertificatesResponse_position - Undocumented member.

$sel:httpStatus:GetClientCertificatesResponse', getClientCertificatesResponse_httpStatus - The response's http status code.

Response Lenses

getClientCertificatesResponse_items :: Lens' GetClientCertificatesResponse (Maybe [ClientCertificate])

The current page of elements from this collection.

getClientCertificatesResponse_position :: Lens' GetClientCertificatesResponse (Maybe Text)

Undocumented member.

getClientCertificatesResponse_httpStatus :: Lens' GetClientCertificatesResponse Int

The response's http status code.