Amazonka.APIGateway.GetRequestValidator

Description

Gets a RequestValidator of a given RestApi.

Creating a Request

data GetRequestValidator

Gets a RequestValidator of a given RestApi.

See: newGetRequestValidator smart constructor.

Constructors

GetRequestValidator'

Fields

restApiId :: Text

Required

The string identifier of the associated RestApi.

requestValidatorId :: Text

Required

The identifier of the RequestValidator to be retrieved.

newGetRequestValidator

Arguments

:: Text
$sel:restApiId:GetRequestValidator'

-> Text
$sel:requestValidatorId:GetRequestValidator'

-> GetRequestValidator

Create a value of GetRequestValidator with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:restApiId:GetRequestValidator', getRequestValidator_restApiId - [Required] The string identifier of the associated RestApi.

$sel:requestValidatorId:GetRequestValidator', getRequestValidator_requestValidatorId - [Required] The identifier of the RequestValidator to be retrieved.

Request Lenses

getRequestValidator_restApiId :: Lens' GetRequestValidator Text

Required

The string identifier of the associated RestApi.

getRequestValidator_requestValidatorId :: Lens' GetRequestValidator Text

Required

The identifier of the RequestValidator to be retrieved.

Destructuring the Response

data RequestValidator

A set of validation rules for incoming Method requests.

In OpenAPI, a RequestValidator of an API is defined by the x-amazon-apigateway-request-validators.requestValidator object. It the referenced using the x-amazon-apigateway-request-validator property.

Enable Basic Request Validation in API Gateway

See: newRequestValidator smart constructor.

Constructors

RequestValidator'

Fields

validateRequestParameters :: Maybe Bool

A Boolean flag to indicate whether to validate request parameters (true) or not (false).

name :: Maybe Text

The name of this RequestValidator

validateRequestBody :: Maybe Bool

A Boolean flag to indicate whether to validate a request body according to the configured Model schema.

id :: Maybe Text

The identifier of this RequestValidator.

newRequestValidator :: RequestValidator

Create a value of RequestValidator with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:validateRequestParameters:RequestValidator', requestValidator_validateRequestParameters - A Boolean flag to indicate whether to validate request parameters (true) or not (false).

$sel:name:RequestValidator', requestValidator_name - The name of this RequestValidator

$sel:validateRequestBody:RequestValidator', requestValidator_validateRequestBody - A Boolean flag to indicate whether to validate a request body according to the configured Model schema.

$sel:id:RequestValidator', requestValidator_id - The identifier of this RequestValidator.

Response Lenses

requestValidator_validateRequestParameters :: Lens' RequestValidator (Maybe Bool)

A Boolean flag to indicate whether to validate request parameters (true) or not (false).

requestValidator_name :: Lens' RequestValidator (Maybe Text)

The name of this RequestValidator

requestValidator_validateRequestBody :: Lens' RequestValidator (Maybe Bool)

A Boolean flag to indicate whether to validate a request body according to the configured Model schema.

requestValidator_id :: Lens' RequestValidator (Maybe Text)

The identifier of this RequestValidator.