Amazonka.APIGateway.GetResources

Description

Lists information about a collection of Resource resources.

This operation returns paginated results.

Creating a Request

data GetResources

Request to list information about a collection of resources.

See: newGetResources smart constructor.

Constructors

GetResources'

Fields

embed :: Maybe [Text]

A query parameter used to retrieve the specified resources embedded in the returned Resources resource in the response. This embed parameter value is a list of comma-separated strings. Currently, the request supports only retrieval of the embedded Method resources this way. The query parameter value must be a single-valued list and contain the "methods" string. For example, GET /restapis/{restapi_id}/resources?embed=methods.

limit :: Maybe Int

The maximum number of returned results per page. The default value is 25 and the maximum value is 500.

position :: Maybe Text

The current pagination position in the paged result set.

restApiId :: Text

Required

The string identifier of the associated RestApi.

newGetResources

Arguments

:: Text
$sel:restApiId:GetResources'

-> GetResources

Create a value of GetResources with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:embed:GetResources', getResources_embed - A query parameter used to retrieve the specified resources embedded in the returned Resources resource in the response. This embed parameter value is a list of comma-separated strings. Currently, the request supports only retrieval of the embedded Method resources this way. The query parameter value must be a single-valued list and contain the "methods" string. For example, GET /restapis/{restapi_id}/resources?embed=methods.

$sel:limit:GetResources', getResources_limit - The maximum number of returned results per page. The default value is 25 and the maximum value is 500.

$sel:position:GetResources', getResources_position - The current pagination position in the paged result set.

$sel:restApiId:GetResources', getResources_restApiId - [Required] The string identifier of the associated RestApi.

Request Lenses

getResources_embed :: Lens' GetResources (Maybe [Text])

A query parameter used to retrieve the specified resources embedded in the returned Resources resource in the response. This embed parameter value is a list of comma-separated strings. Currently, the request supports only retrieval of the embedded Method resources this way. The query parameter value must be a single-valued list and contain the "methods" string. For example, GET /restapis/{restapi_id}/resources?embed=methods.

getResources_limit :: Lens' GetResources (Maybe Int)

The maximum number of returned results per page. The default value is 25 and the maximum value is 500.

getResources_position :: Lens' GetResources (Maybe Text)

The current pagination position in the paged result set.

getResources_restApiId :: Lens' GetResources Text

Required

The string identifier of the associated RestApi.

Destructuring the Response

data GetResourcesResponse

Represents a collection of Resource resources.

Create an API

See: newGetResourcesResponse smart constructor.

Constructors

GetResourcesResponse'

Fields

items :: Maybe [Resource]

The current page of elements from this collection.

position :: Maybe Text

httpStatus :: Int

The response's http status code.

newGetResourcesResponse

Arguments

:: Int
$sel:httpStatus:GetResourcesResponse'

-> GetResourcesResponse

Create a value of GetResourcesResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:items:GetResourcesResponse', getResourcesResponse_items - The current page of elements from this collection.

$sel:position:GetResources', getResourcesResponse_position - Undocumented member.

$sel:httpStatus:GetResourcesResponse', getResourcesResponse_httpStatus - The response's http status code.

Response Lenses

getResourcesResponse_items :: Lens' GetResourcesResponse (Maybe [Resource])

The current page of elements from this collection.

getResourcesResponse_position :: Lens' GetResourcesResponse (Maybe Text)

Undocumented member.

getResourcesResponse_httpStatus :: Lens' GetResourcesResponse Int

The response's http status code.