Amazonka.APIGateway.GetSdk

Description

Generates a client SDK for a RestApi and Stage.

Creating a Request

data GetSdk

Request a new generated client SDK for a RestApi and Stage.

See: newGetSdk smart constructor.

Constructors

GetSdk'

Fields

parameters :: Maybe (HashMap Text Text)

A string-to-string key-value map of query parameters sdkType-dependent properties of the SDK. For sdkType of objectivec or swift, a parameter named classPrefix is required. For sdkType of android, parameters named groupId, artifactId, artifactVersion, and invokerPackage are required. For sdkType of java, parameters named serviceName and javaPackageName are required.

restApiId :: Text

Required

The string identifier of the associated RestApi.

stageName :: Text

Required

The name of the Stage that the SDK will use.

sdkType :: Text

Required

The language for the generated SDK. Currently java, javascript, android, objectivec (for iOS), swift (for iOS), and ruby are supported.

newGetSdk

Arguments

:: Text
$sel:restApiId:GetSdk'

-> Text
$sel:stageName:GetSdk'

-> Text
$sel:sdkType:GetSdk'

-> GetSdk

Create a value of GetSdk with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:parameters:GetSdk', getSdk_parameters - A string-to-string key-value map of query parameters sdkType-dependent properties of the SDK. For sdkType of objectivec or swift, a parameter named classPrefix is required. For sdkType of android, parameters named groupId, artifactId, artifactVersion, and invokerPackage are required. For sdkType of java, parameters named serviceName and javaPackageName are required.

$sel:restApiId:GetSdk', getSdk_restApiId - [Required] The string identifier of the associated RestApi.

$sel:stageName:GetSdk', getSdk_stageName - [Required] The name of the Stage that the SDK will use.

$sel:sdkType:GetSdk', getSdk_sdkType - [Required] The language for the generated SDK. Currently java, javascript, android, objectivec (for iOS), swift (for iOS), and ruby are supported.

Request Lenses

getSdk_parameters :: Lens' GetSdk (Maybe (HashMap Text Text))

A string-to-string key-value map of query parameters sdkType-dependent properties of the SDK. For sdkType of objectivec or swift, a parameter named classPrefix is required. For sdkType of android, parameters named groupId, artifactId, artifactVersion, and invokerPackage are required. For sdkType of java, parameters named serviceName and javaPackageName are required.

getSdk_restApiId :: Lens' GetSdk Text

Required

The string identifier of the associated RestApi.

getSdk_stageName :: Lens' GetSdk Text

Required

The name of the Stage that the SDK will use.

getSdk_sdkType :: Lens' GetSdk Text

Required

The language for the generated SDK. Currently java, javascript, android, objectivec (for iOS), swift (for iOS), and ruby are supported.

Destructuring the Response

data GetSdkResponse

The binary blob response to GetSdk, which contains the generated SDK.

See: newGetSdkResponse smart constructor.

Constructors

GetSdkResponse'

Fields

body :: Maybe ByteString

The binary blob response to GetSdk, which contains the generated SDK.

contentDisposition :: Maybe Text

The content-disposition header value in the HTTP response.

contentType :: Maybe Text

The content-type header value in the HTTP response.

httpStatus :: Int

The response's http status code.

newGetSdkResponse

Arguments

:: Int
$sel:httpStatus:GetSdkResponse'

-> GetSdkResponse

Create a value of GetSdkResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:body:GetSdkResponse', getSdkResponse_body - The binary blob response to GetSdk, which contains the generated SDK.

$sel:contentDisposition:GetSdkResponse', getSdkResponse_contentDisposition - The content-disposition header value in the HTTP response.

$sel:contentType:GetSdkResponse', getSdkResponse_contentType - The content-type header value in the HTTP response.

$sel:httpStatus:GetSdkResponse', getSdkResponse_httpStatus - The response's http status code.

Response Lenses

getSdkResponse_body :: Lens' GetSdkResponse (Maybe ByteString)

The binary blob response to GetSdk, which contains the generated SDK.

getSdkResponse_contentDisposition :: Lens' GetSdkResponse (Maybe Text)

The content-disposition header value in the HTTP response.

getSdkResponse_contentType :: Lens' GetSdkResponse (Maybe Text)

The content-type header value in the HTTP response.

getSdkResponse_httpStatus :: Lens' GetSdkResponse Int

The response's http status code.