Amazonka.APIGateway.GetSdkTypes

Description

- | Undocumented operation.

This operation returns paginated results.

Creating a Request

data GetSdkTypes

Get the SdkTypes collection.

See: newGetSdkTypes smart constructor.

Constructors

GetSdkTypes'

Fields

limit :: Maybe Int

The maximum number of returned results per page. The default value is 25 and the maximum value is 500.

position :: Maybe Text

The current pagination position in the paged result set.

newGetSdkTypes :: GetSdkTypes

Create a value of GetSdkTypes with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:limit:GetSdkTypes', getSdkTypes_limit - The maximum number of returned results per page. The default value is 25 and the maximum value is 500.

$sel:position:GetSdkTypes', getSdkTypes_position - The current pagination position in the paged result set.

Request Lenses

getSdkTypes_limit :: Lens' GetSdkTypes (Maybe Int)

The maximum number of returned results per page. The default value is 25 and the maximum value is 500.

getSdkTypes_position :: Lens' GetSdkTypes (Maybe Text)

The current pagination position in the paged result set.

Destructuring the Response

data GetSdkTypesResponse

The collection of SdkType instances.

See: newGetSdkTypesResponse smart constructor.

Constructors

GetSdkTypesResponse'

Fields

items :: Maybe [SdkType]

The current page of elements from this collection.

position :: Maybe Text

httpStatus :: Int

The response's http status code.

newGetSdkTypesResponse

Arguments

:: Int
$sel:httpStatus:GetSdkTypesResponse'

-> GetSdkTypesResponse

Create a value of GetSdkTypesResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:items:GetSdkTypesResponse', getSdkTypesResponse_items - The current page of elements from this collection.

$sel:position:GetSdkTypes', getSdkTypesResponse_position - Undocumented member.

$sel:httpStatus:GetSdkTypesResponse', getSdkTypesResponse_httpStatus - The response's http status code.

Response Lenses

getSdkTypesResponse_items :: Lens' GetSdkTypesResponse (Maybe [SdkType])

The current page of elements from this collection.

getSdkTypesResponse_position :: Lens' GetSdkTypesResponse (Maybe Text)

Undocumented member.

getSdkTypesResponse_httpStatus :: Lens' GetSdkTypesResponse Int

The response's http status code.