Amazonka.APIGateway.PutMethodResponse

Description

Adds a MethodResponse to an existing Method resource.

Creating a Request

data PutMethodResponse

Request to add a MethodResponse to an existing Method resource.

See: newPutMethodResponse smart constructor.

Constructors

PutMethodResponse'

Fields

responseModels :: Maybe (HashMap Text Text)

Specifies the Model resources used for the response's content type. Response models are represented as a key/value map, with a content type as the key and a Model name as the value.

responseParameters :: Maybe (HashMap Text Bool)

A key-value map specifying required or optional response parameters that API Gateway can send back to the caller. A key defines a method response header name and the associated value is a Boolean flag indicating whether the method response parameter is required or not. The method response header names must match the pattern of method.response.header.{name}, where name is a valid and unique header name. The response parameter names defined here are available in the integration response to be mapped from an integration response header expressed in integration.response.header.{name}, a static value enclosed within a pair of single quotes (e.g., 'application/json'), or a JSON expression from the back-end response payload in the form of integration.response.body.{JSON-expression}, where JSON-expression is a valid JSON expression without the $ prefix.)

restApiId :: Text

Required

The string identifier of the associated RestApi.

resourceId :: Text

Required

The Resource identifier for the Method resource.

httpMethod :: Text

Required

The HTTP verb of the Method resource.

statusCode :: Text

Required

The method response's status code.

newPutMethodResponse

Arguments

:: Text
$sel:restApiId:PutMethodResponse'

-> Text
$sel:resourceId:PutMethodResponse'

-> Text
$sel:httpMethod:PutMethodResponse'

-> Text
$sel:statusCode:PutMethodResponse'

-> PutMethodResponse

Create a value of PutMethodResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:responseModels:PutMethodResponse', putMethodResponse_responseModels - Specifies the Model resources used for the response's content type. Response models are represented as a key/value map, with a content type as the key and a Model name as the value.

$sel:responseParameters:PutMethodResponse', putMethodResponse_responseParameters - A key-value map specifying required or optional response parameters that API Gateway can send back to the caller. A key defines a method response header name and the associated value is a Boolean flag indicating whether the method response parameter is required or not. The method response header names must match the pattern of method.response.header.{name}, where name is a valid and unique header name. The response parameter names defined here are available in the integration response to be mapped from an integration response header expressed in integration.response.header.{name}, a static value enclosed within a pair of single quotes (e.g., 'application/json'), or a JSON expression from the back-end response payload in the form of integration.response.body.{JSON-expression}, where JSON-expression is a valid JSON expression without the $ prefix.)

$sel:restApiId:PutMethodResponse', putMethodResponse_restApiId - [Required] The string identifier of the associated RestApi.

$sel:resourceId:PutMethodResponse', putMethodResponse_resourceId - [Required] The Resource identifier for the Method resource.

$sel:httpMethod:PutMethodResponse', putMethodResponse_httpMethod - [Required] The HTTP verb of the Method resource.

$sel:statusCode:PutMethodResponse', putMethodResponse_statusCode - [Required] The method response's status code.

Request Lenses

putMethodResponse_responseModels :: Lens' PutMethodResponse (Maybe (HashMap Text Text))

Specifies the Model resources used for the response's content type. Response models are represented as a key/value map, with a content type as the key and a Model name as the value.

putMethodResponse_responseParameters :: Lens' PutMethodResponse (Maybe (HashMap Text Bool))

A key-value map specifying required or optional response parameters that API Gateway can send back to the caller. A key defines a method response header name and the associated value is a Boolean flag indicating whether the method response parameter is required or not. The method response header names must match the pattern of method.response.header.{name}, where name is a valid and unique header name. The response parameter names defined here are available in the integration response to be mapped from an integration response header expressed in integration.response.header.{name}, a static value enclosed within a pair of single quotes (e.g., 'application/json'), or a JSON expression from the back-end response payload in the form of integration.response.body.{JSON-expression}, where JSON-expression is a valid JSON expression without the $ prefix.)

putMethodResponse_restApiId :: Lens' PutMethodResponse Text

Required

The string identifier of the associated RestApi.

putMethodResponse_resourceId :: Lens' PutMethodResponse Text

Required

The Resource identifier for the Method resource.

putMethodResponse_httpMethod :: Lens' PutMethodResponse Text

Required

The HTTP verb of the Method resource.

putMethodResponse_statusCode :: Lens' PutMethodResponse Text

Required

The method response's status code.

Destructuring the Response

data MethodResponse

Represents a method response of a given HTTP status code returned to the client. The method response is passed from the back end through the associated integration response that can be transformed using a mapping template.

Example: A MethodResponse instance of an API

Request

The example request retrieves a MethodResponse of the 200 status code.

GET /restapis/fugvjdxtri/resources/3kzxbg5sa2/methods/GET/responses/200 HTTP/1.1 Content-Type: application/json Host: apigateway.us-east-1.amazonaws.com X-Amz-Date: 20160603T222952Z Authorization: AWS4-HMAC-SHA256 Credential={access_key_ID}/20160603/us-east-1/apigateway/aws4_request, SignedHeaders=content-type;host;x-amz-date, Signature={sig4_hash}

Response

The successful response returns 200 OK status and a payload as follows:

{ "_links": { "curies": { "href": "https://docs.aws.amazon.com/apigateway/latest/developerguide/restapi-method-response-{rel}.html", "name": "methodresponse", "templated": true }, "self": { "href": "/restapis/fugvjdxtri/resources/3kzxbg5sa2/methods/GET/responses/200", "title": "200" }, "methodresponse:delete": { "href": "/restapis/fugvjdxtri/resources/3kzxbg5sa2/methods/GET/responses/200" }, "methodresponse:update": { "href": "/restapis/fugvjdxtri/resources/3kzxbg5sa2/methods/GET/responses/200" } }, "responseModels": { "application/json": "Empty" }, "responseParameters": { "method.response.header.Content-Type": false }, "statusCode": "200" }

Method, IntegrationResponse, Integration Creating an API

See: newMethodResponse smart constructor.

Constructors

MethodResponse'

Fields

responseModels :: Maybe (HashMap Text Text)

Specifies the Model resources used for the response's content-type. Response models are represented as a key/value map, with a content-type as the key and a Model name as the value.

statusCode :: Maybe Text

The method response's status code.

responseParameters :: Maybe (HashMap Text Bool)

A key-value map specifying required or optional response parameters that API Gateway can send back to the caller. A key defines a method response header and the value specifies whether the associated method response header is required or not. The expression of the key must match the pattern method.response.header.{name}, where name is a valid and unique header name. API Gateway passes certain integration response data to the method response headers specified here according to the mapping you prescribe in the API's IntegrationResponse. The integration response data that can be mapped include an integration response header expressed in integration.response.header.{name}, a static value enclosed within a pair of single quotes (e.g., 'application/json'), or a JSON expression from the back-end response payload in the form of integration.response.body.{JSON-expression}, where JSON-expression is a valid JSON expression without the $ prefix.)

newMethodResponse :: MethodResponse

Create a value of MethodResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:responseModels:MethodResponse', methodResponse_responseModels - Specifies the Model resources used for the response's content-type. Response models are represented as a key/value map, with a content-type as the key and a Model name as the value.

$sel:statusCode:MethodResponse', methodResponse_statusCode - The method response's status code.

$sel:responseParameters:MethodResponse', methodResponse_responseParameters - A key-value map specifying required or optional response parameters that API Gateway can send back to the caller. A key defines a method response header and the value specifies whether the associated method response header is required or not. The expression of the key must match the pattern method.response.header.{name}, where name is a valid and unique header name. API Gateway passes certain integration response data to the method response headers specified here according to the mapping you prescribe in the API's IntegrationResponse. The integration response data that can be mapped include an integration response header expressed in integration.response.header.{name}, a static value enclosed within a pair of single quotes (e.g., 'application/json'), or a JSON expression from the back-end response payload in the form of integration.response.body.{JSON-expression}, where JSON-expression is a valid JSON expression without the $ prefix.)

Response Lenses

methodResponse_responseModels :: Lens' MethodResponse (Maybe (HashMap Text Text))

Specifies the Model resources used for the response's content-type. Response models are represented as a key/value map, with a content-type as the key and a Model name as the value.

methodResponse_statusCode :: Lens' MethodResponse (Maybe Text)

The method response's status code.

methodResponse_responseParameters :: Lens' MethodResponse (Maybe (HashMap Text Bool))

A key-value map specifying required or optional response parameters that API Gateway can send back to the caller. A key defines a method response header and the value specifies whether the associated method response header is required or not. The expression of the key must match the pattern method.response.header.{name}, where name is a valid and unique header name. API Gateway passes certain integration response data to the method response headers specified here according to the mapping you prescribe in the API's IntegrationResponse. The integration response data that can be mapped include an integration response header expressed in integration.response.header.{name}, a static value enclosed within a pair of single quotes (e.g., 'application/json'), or a JSON expression from the back-end response payload in the form of integration.response.body.{JSON-expression}, where JSON-expression is a valid JSON expression without the $ prefix.)