Amazonka.APIGateway.TestInvokeMethod

Description

Simulate the execution of a Method in your RestApi with headers, parameters, and an incoming request body.

Creating a Request

data TestInvokeMethod

Make a request to simulate the execution of a Method.

See: newTestInvokeMethod smart constructor.

Constructors

TestInvokeMethod'

Fields

pathWithQueryString :: Maybe Text

The URI path, including query string, of the simulated invocation request. Use this to specify path parameters and query string parameters.

body :: Maybe Text

The simulated request body of an incoming invocation request.

clientCertificateId :: Maybe Text

A ClientCertificate identifier to use in the test invocation. API Gateway will use the certificate when making the HTTPS request to the defined back-end endpoint.

stageVariables :: Maybe (HashMap Text Text)

A key-value map of stage variables to simulate an invocation on a deployed Stage.

headers :: Maybe (HashMap Text Text)

A key-value map of headers to simulate an incoming invocation request.

multiValueHeaders :: Maybe (HashMap Text [Text])

The headers as a map from string to list of values to simulate an incoming invocation request.

restApiId :: Text

Required

The string identifier of the associated RestApi.

resourceId :: Text

Required

Specifies a test invoke method request's resource ID.

httpMethod :: Text

Required

Specifies a test invoke method request's HTTP method.

newTestInvokeMethod

Arguments

:: Text
$sel:restApiId:TestInvokeMethod'

-> Text
$sel:resourceId:TestInvokeMethod'

-> Text
$sel:httpMethod:TestInvokeMethod'

-> TestInvokeMethod

Create a value of TestInvokeMethod with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:pathWithQueryString:TestInvokeMethod', testInvokeMethod_pathWithQueryString - The URI path, including query string, of the simulated invocation request. Use this to specify path parameters and query string parameters.

$sel:body:TestInvokeMethod', testInvokeMethod_body - The simulated request body of an incoming invocation request.

$sel:clientCertificateId:TestInvokeMethod', testInvokeMethod_clientCertificateId - A ClientCertificate identifier to use in the test invocation. API Gateway will use the certificate when making the HTTPS request to the defined back-end endpoint.

$sel:stageVariables:TestInvokeMethod', testInvokeMethod_stageVariables - A key-value map of stage variables to simulate an invocation on a deployed Stage.

$sel:headers:TestInvokeMethod', testInvokeMethod_headers - A key-value map of headers to simulate an incoming invocation request.

$sel:multiValueHeaders:TestInvokeMethod', testInvokeMethod_multiValueHeaders - The headers as a map from string to list of values to simulate an incoming invocation request.

$sel:restApiId:TestInvokeMethod', testInvokeMethod_restApiId - [Required] The string identifier of the associated RestApi.

$sel:resourceId:TestInvokeMethod', testInvokeMethod_resourceId - [Required] Specifies a test invoke method request's resource ID.

$sel:httpMethod:TestInvokeMethod', testInvokeMethod_httpMethod - [Required] Specifies a test invoke method request's HTTP method.

Request Lenses

testInvokeMethod_pathWithQueryString :: Lens' TestInvokeMethod (Maybe Text)

The URI path, including query string, of the simulated invocation request. Use this to specify path parameters and query string parameters.

testInvokeMethod_body :: Lens' TestInvokeMethod (Maybe Text)

The simulated request body of an incoming invocation request.

testInvokeMethod_clientCertificateId :: Lens' TestInvokeMethod (Maybe Text)

A ClientCertificate identifier to use in the test invocation. API Gateway will use the certificate when making the HTTPS request to the defined back-end endpoint.

testInvokeMethod_stageVariables :: Lens' TestInvokeMethod (Maybe (HashMap Text Text))

A key-value map of stage variables to simulate an invocation on a deployed Stage.

testInvokeMethod_headers :: Lens' TestInvokeMethod (Maybe (HashMap Text Text))

A key-value map of headers to simulate an incoming invocation request.

testInvokeMethod_multiValueHeaders :: Lens' TestInvokeMethod (Maybe (HashMap Text [Text]))

The headers as a map from string to list of values to simulate an incoming invocation request.

testInvokeMethod_restApiId :: Lens' TestInvokeMethod Text

Required

The string identifier of the associated RestApi.

testInvokeMethod_resourceId :: Lens' TestInvokeMethod Text

Required

Specifies a test invoke method request's resource ID.

testInvokeMethod_httpMethod :: Lens' TestInvokeMethod Text

Required

Specifies a test invoke method request's HTTP method.

Destructuring the Response

data TestInvokeMethodResponse

Represents the response of the test invoke request in the HTTP method.

Test API using the API Gateway console

See: newTestInvokeMethodResponse smart constructor.

Constructors

TestInvokeMethodResponse'

Fields

log :: Maybe Text

The API Gateway execution log for the test invoke request.

status :: Maybe Int

The HTTP status code.

body :: Maybe Text

The body of the HTTP response.

latency :: Maybe Integer

The execution latency of the test invoke request.

headers :: Maybe (HashMap Text Text)

The headers of the HTTP response.

multiValueHeaders :: Maybe (HashMap Text [Text])

The headers of the HTTP response as a map from string to list of values.

httpStatus :: Int

The response's http status code.

newTestInvokeMethodResponse

Arguments

:: Int
$sel:httpStatus:TestInvokeMethodResponse'

-> TestInvokeMethodResponse

Create a value of TestInvokeMethodResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:log:TestInvokeMethodResponse', testInvokeMethodResponse_log - The API Gateway execution log for the test invoke request.

$sel:status:TestInvokeMethodResponse', testInvokeMethodResponse_status - The HTTP status code.

$sel:body:TestInvokeMethod', testInvokeMethodResponse_body - The body of the HTTP response.

$sel:latency:TestInvokeMethodResponse', testInvokeMethodResponse_latency - The execution latency of the test invoke request.

$sel:headers:TestInvokeMethod', testInvokeMethodResponse_headers - The headers of the HTTP response.

$sel:multiValueHeaders:TestInvokeMethod', testInvokeMethodResponse_multiValueHeaders - The headers of the HTTP response as a map from string to list of values.

$sel:httpStatus:TestInvokeMethodResponse', testInvokeMethodResponse_httpStatus - The response's http status code.

Response Lenses

testInvokeMethodResponse_log :: Lens' TestInvokeMethodResponse (Maybe Text)

The API Gateway execution log for the test invoke request.

testInvokeMethodResponse_status :: Lens' TestInvokeMethodResponse (Maybe Int)

The HTTP status code.

testInvokeMethodResponse_body :: Lens' TestInvokeMethodResponse (Maybe Text)

The body of the HTTP response.

testInvokeMethodResponse_latency :: Lens' TestInvokeMethodResponse (Maybe Integer)

The execution latency of the test invoke request.

testInvokeMethodResponse_headers :: Lens' TestInvokeMethodResponse (Maybe (HashMap Text Text))

The headers of the HTTP response.

testInvokeMethodResponse_multiValueHeaders :: Lens' TestInvokeMethodResponse (Maybe (HashMap Text [Text]))

The headers of the HTTP response as a map from string to list of values.

testInvokeMethodResponse_httpStatus :: Lens' TestInvokeMethodResponse Int

The response's http status code.