Amazonka.APIGateway.Types.BasePathMapping

Description

Documentation

data BasePathMapping

Represents the base path that callers of the API must provide as part of the URL after the domain name.

A custom domain name plus a BasePathMapping specification identifies a deployed RestApi in a given stage of the owner Account.

Use Custom Domain Names

See: newBasePathMapping smart constructor.

Constructors

BasePathMapping'

Fields

stage :: Maybe Text

The name of the associated stage.

basePath :: Maybe Text

The base path name that callers of the API must provide as part of the URL after the domain name.

restApiId :: Maybe Text

The string identifier of the associated RestApi.

newBasePathMapping :: BasePathMapping

Create a value of BasePathMapping with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:stage:BasePathMapping', basePathMapping_stage - The name of the associated stage.

$sel:basePath:BasePathMapping', basePathMapping_basePath - The base path name that callers of the API must provide as part of the URL after the domain name.

$sel:restApiId:BasePathMapping', basePathMapping_restApiId - The string identifier of the associated RestApi.

basePathMapping_stage :: Lens' BasePathMapping (Maybe Text)

The name of the associated stage.

basePathMapping_basePath :: Lens' BasePathMapping (Maybe Text)

The base path name that callers of the API must provide as part of the URL after the domain name.

basePathMapping_restApiId :: Lens' BasePathMapping (Maybe Text)

The string identifier of the associated RestApi.