Amazonka.APIGatewayManagementAPI

Description

Derived from API version 2018-11-29 of the AWS service descriptions, licensed under Apache 2.0.

The Amazon API Gateway Management API allows you to directly manage runtime aspects of your deployed APIs. To use it, you must explicitly set the SDK's endpoint to point to the endpoint of your deployed API. The endpoint will be of the form https://{api-id}.execute-api.{region}.amazonaws.com/{stage}, or will be the endpoint corresponding to your API's custom domain and base path, if applicable.

Service Configuration

defaultService :: Service

API version 2018-11-29 of the Amazon ApiGatewayManagementApi SDK configuration.

Errors

Error matchers are designed for use with the functions provided by Control.Exception.Lens. This allows catching (and rethrowing) service specific errors returned by APIGatewayManagementAPI.

PayloadTooLargeException

_PayloadTooLargeException :: AsError a => Getting (First ServiceError) a ServiceError

The data has exceeded the maximum size allowed.

ForbiddenException

_ForbiddenException :: AsError a => Getting (First ServiceError) a ServiceError

The caller is not authorized to invoke this operation.

GoneException

_GoneException :: AsError a => Getting (First ServiceError) a ServiceError

The connection with the provided id no longer exists.

LimitExceededException

_LimitExceededException :: AsError a => Getting (First ServiceError) a ServiceError

The client is sending more than the allowed number of requests per unit of time or the WebSocket client side buffer is full.

Waiters

Waiters poll by repeatedly sending a request until some remote success condition configured by the Wait specification is fulfilled. The Wait specification determines how many attempts should be made, in addition to delay and retry strategies.

Operations

Some AWS operations return results that are incomplete and require subsequent requests in order to obtain the entire result set. The process of sending subsequent requests to continue where a previous request left off is called pagination. For example, the ListObjects operation of Amazon S3 returns up to 1000 objects at a time, and you must send subsequent requests with the appropriate Marker in order to retrieve the next page of results.

Operations that have an AWSPager instance can transparently perform subsequent requests, correctly setting Markers and other request facets to iterate through the entire result set of a truncated API operation. Operations which support this have an additional note in the documentation.

Many operations have the ability to filter results on the server side. See the individual operation parameters for details.

DeleteConnection

data DeleteConnection

See: newDeleteConnection smart constructor.

Constructors

DeleteConnection' Text

newDeleteConnection

Arguments

:: Text
$sel:connectionId:DeleteConnection'

-> DeleteConnection

Create a value of DeleteConnection with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:connectionId:DeleteConnection', deleteConnection_connectionId - Undocumented member.

data DeleteConnectionResponse

See: newDeleteConnectionResponse smart constructor.

Constructors

DeleteConnectionResponse'

newDeleteConnectionResponse :: DeleteConnectionResponse

Create a value of DeleteConnectionResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

GetConnection

data GetConnection

See: newGetConnection smart constructor.

Constructors

GetConnection' Text

newGetConnection

Arguments

:: Text
$sel:connectionId:GetConnection'

-> GetConnection

Create a value of GetConnection with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:connectionId:GetConnection', getConnection_connectionId - Undocumented member.

data GetConnectionResponse

See: newGetConnectionResponse smart constructor.

Constructors

GetConnectionResponse' (Maybe POSIX) (Maybe POSIX) (Maybe Identity) Int

newGetConnectionResponse

Arguments

:: Int
$sel:httpStatus:GetConnectionResponse'

-> GetConnectionResponse

Create a value of GetConnectionResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:connectedAt:GetConnectionResponse', getConnectionResponse_connectedAt - The time in ISO 8601 format for when the connection was established.

$sel:lastActiveAt:GetConnectionResponse', getConnectionResponse_lastActiveAt - The time in ISO 8601 format for when the connection was last active.

$sel:identity:GetConnectionResponse', getConnectionResponse_identity - Undocumented member.

$sel:httpStatus:GetConnectionResponse', getConnectionResponse_httpStatus - The response's http status code.

PostToConnection

data PostToConnection

See: newPostToConnection smart constructor.

Constructors

PostToConnection' Text ByteString

newPostToConnection

Arguments

:: Text
$sel:connectionId:PostToConnection'

-> ByteString
$sel:data':PostToConnection'

-> PostToConnection

Create a value of PostToConnection with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:connectionId:PostToConnection', postToConnection_connectionId - The identifier of the connection that a specific client is using.

$sel:data':PostToConnection', postToConnection_data - The data to be sent to the client specified by its connection id.

data PostToConnectionResponse

See: newPostToConnectionResponse smart constructor.

Constructors

PostToConnectionResponse'

newPostToConnectionResponse :: PostToConnectionResponse

Create a value of PostToConnectionResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

Types

Identity

data Identity

See: newIdentity smart constructor.

Constructors

Identity' Text Text

newIdentity

Arguments

:: Text
$sel:sourceIp:Identity'

-> Text
$sel:userAgent:Identity'

-> Identity

Create a value of Identity with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:sourceIp:Identity', identity_sourceIp - The source IP address of the TCP connection making the request to API Gateway.

$sel:userAgent:Identity', identity_userAgent - The User Agent of the API caller.