Amazonka.AWSHealth.DescribeAffectedEntities

Description

Returns a list of entities that have been affected by the specified events, based on the specified filter criteria. Entities can refer to individual customer resources, groups of customer resources, or any other construct, depending on the AWS service. Events that have impact beyond that of the affected entities, or where the extent of impact is unknown, include at least one entity indicating this.

At least one event ARN is required. Results are sorted by the lastUpdatedTime of the entity, starting with the most recent.

This API operation uses pagination. Specify the nextToken parameter in the next request to return more results.

This operation supports resource-level permissions. You can use this operation to allow or deny access to specific AWS Health events. For more information, see Resource- and action-based conditions in the AWS Health User Guide.

This operation returns paginated results.

Creating a Request

data DescribeAffectedEntities

See: newDescribeAffectedEntities smart constructor.

Constructors

DescribeAffectedEntities'

Fields

locale :: Maybe Text

The locale (language) to return information in. English (en) is the default and the only supported value at this time.

nextToken :: Maybe Text

If the results of a search are large, only a portion of the results are returned, and a nextToken pagination token is returned in the response. To retrieve the next batch of results, reissue the search request and include the returned token. When all results have been returned, the response does not contain a pagination token value.

maxResults :: Maybe Natural

The maximum number of items to return in one batch, between 10 and 100, inclusive.

filter' :: EntityFilter

Values to narrow the results returned. At least one event ARN is required.

newDescribeAffectedEntities

Arguments

:: EntityFilter
$sel:filter':DescribeAffectedEntities'

-> DescribeAffectedEntities

Create a value of DescribeAffectedEntities with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:locale:DescribeAffectedEntities', describeAffectedEntities_locale - The locale (language) to return information in. English (en) is the default and the only supported value at this time.

$sel:nextToken:DescribeAffectedEntities', describeAffectedEntities_nextToken - If the results of a search are large, only a portion of the results are returned, and a nextToken pagination token is returned in the response. To retrieve the next batch of results, reissue the search request and include the returned token. When all results have been returned, the response does not contain a pagination token value.

$sel:maxResults:DescribeAffectedEntities', describeAffectedEntities_maxResults - The maximum number of items to return in one batch, between 10 and 100, inclusive.

$sel:filter':DescribeAffectedEntities', describeAffectedEntities_filter - Values to narrow the results returned. At least one event ARN is required.

Request Lenses

describeAffectedEntities_locale :: Lens' DescribeAffectedEntities (Maybe Text)

The locale (language) to return information in. English (en) is the default and the only supported value at this time.

describeAffectedEntities_nextToken :: Lens' DescribeAffectedEntities (Maybe Text)

If the results of a search are large, only a portion of the results are returned, and a nextToken pagination token is returned in the response. To retrieve the next batch of results, reissue the search request and include the returned token. When all results have been returned, the response does not contain a pagination token value.

describeAffectedEntities_maxResults :: Lens' DescribeAffectedEntities (Maybe Natural)

The maximum number of items to return in one batch, between 10 and 100, inclusive.

describeAffectedEntities_filter :: Lens' DescribeAffectedEntities EntityFilter

Values to narrow the results returned. At least one event ARN is required.

Destructuring the Response

data DescribeAffectedEntitiesResponse

See: newDescribeAffectedEntitiesResponse smart constructor.

Constructors

DescribeAffectedEntitiesResponse'

Fields

entities :: Maybe [AffectedEntity]

The entities that match the filter criteria.

nextToken :: Maybe Text

If the results of a search are large, only a portion of the results are returned, and a nextToken pagination token is returned in the response. To retrieve the next batch of results, reissue the search request and include the returned token. When all results have been returned, the response does not contain a pagination token value.

httpStatus :: Int

The response's http status code.

newDescribeAffectedEntitiesResponse

Arguments

:: Int
$sel:httpStatus:DescribeAffectedEntitiesResponse'

-> DescribeAffectedEntitiesResponse

Create a value of DescribeAffectedEntitiesResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:entities:DescribeAffectedEntitiesResponse', describeAffectedEntitiesResponse_entities - The entities that match the filter criteria.

$sel:nextToken:DescribeAffectedEntities', describeAffectedEntitiesResponse_nextToken - If the results of a search are large, only a portion of the results are returned, and a nextToken pagination token is returned in the response. To retrieve the next batch of results, reissue the search request and include the returned token. When all results have been returned, the response does not contain a pagination token value.

$sel:httpStatus:DescribeAffectedEntitiesResponse', describeAffectedEntitiesResponse_httpStatus - The response's http status code.

Response Lenses

describeAffectedEntitiesResponse_entities :: Lens' DescribeAffectedEntitiesResponse (Maybe [AffectedEntity])

The entities that match the filter criteria.

describeAffectedEntitiesResponse_nextToken :: Lens' DescribeAffectedEntitiesResponse (Maybe Text)

If the results of a search are large, only a portion of the results are returned, and a nextToken pagination token is returned in the response. To retrieve the next batch of results, reissue the search request and include the returned token. When all results have been returned, the response does not contain a pagination token value.

describeAffectedEntitiesResponse_httpStatus :: Lens' DescribeAffectedEntitiesResponse Int

The response's http status code.