Amazonka.AWSHealth.DescribeAffectedEntitiesForOrganization

Description

Returns a list of entities that have been affected by one or more events for one or more accounts in your organization in AWS Organizations, based on the filter criteria. Entities can refer to individual customer resources, groups of customer resources, or any other construct, depending on the AWS service.

At least one event Amazon Resource Name (ARN) and account ID are required. Results are sorted by the lastUpdatedTime of the entity, starting with the most recent.

Before you can call this operation, you must first enable AWS Health to work with AWS Organizations. To do this, call the EnableHealthServiceAccessForOrganization operation from your organization's management account.

This API operation uses pagination. Specify the nextToken parameter in the next request to return more results.

This operation doesn't support resource-level permissions. You can't use this operation to allow or deny access to specific AWS Health events. For more information, see Resource- and action-based conditions in the AWS Health User Guide.

This operation returns paginated results.

Creating a Request

data DescribeAffectedEntitiesForOrganization

See: newDescribeAffectedEntitiesForOrganization smart constructor.

Constructors

DescribeAffectedEntitiesForOrganization'

Fields

locale :: Maybe Text

The locale (language) to return information in. English (en) is the default and the only supported value at this time.

nextToken :: Maybe Text

If the results of a search are large, only a portion of the results are returned, and a nextToken pagination token is returned in the response. To retrieve the next batch of results, reissue the search request and include the returned token. When all results have been returned, the response does not contain a pagination token value.

maxResults :: Maybe Natural

The maximum number of items to return in one batch, between 10 and 100, inclusive.

organizationEntityFilters :: NonEmpty EventAccountFilter

A JSON set of elements including the awsAccountId and the eventArn.

newDescribeAffectedEntitiesForOrganization

Arguments

:: NonEmpty EventAccountFilter
$sel:organizationEntityFilters:DescribeAffectedEntitiesForOrganization'

-> DescribeAffectedEntitiesForOrganization

Create a value of DescribeAffectedEntitiesForOrganization with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:locale:DescribeAffectedEntitiesForOrganization', describeAffectedEntitiesForOrganization_locale - The locale (language) to return information in. English (en) is the default and the only supported value at this time.

$sel:nextToken:DescribeAffectedEntitiesForOrganization', describeAffectedEntitiesForOrganization_nextToken - If the results of a search are large, only a portion of the results are returned, and a nextToken pagination token is returned in the response. To retrieve the next batch of results, reissue the search request and include the returned token. When all results have been returned, the response does not contain a pagination token value.

$sel:maxResults:DescribeAffectedEntitiesForOrganization', describeAffectedEntitiesForOrganization_maxResults - The maximum number of items to return in one batch, between 10 and 100, inclusive.

$sel:organizationEntityFilters:DescribeAffectedEntitiesForOrganization', describeAffectedEntitiesForOrganization_organizationEntityFilters - A JSON set of elements including the awsAccountId and the eventArn.

Request Lenses

describeAffectedEntitiesForOrganization_locale :: Lens' DescribeAffectedEntitiesForOrganization (Maybe Text)

The locale (language) to return information in. English (en) is the default and the only supported value at this time.

describeAffectedEntitiesForOrganization_nextToken :: Lens' DescribeAffectedEntitiesForOrganization (Maybe Text)

If the results of a search are large, only a portion of the results are returned, and a nextToken pagination token is returned in the response. To retrieve the next batch of results, reissue the search request and include the returned token. When all results have been returned, the response does not contain a pagination token value.

describeAffectedEntitiesForOrganization_maxResults :: Lens' DescribeAffectedEntitiesForOrganization (Maybe Natural)

The maximum number of items to return in one batch, between 10 and 100, inclusive.

describeAffectedEntitiesForOrganization_organizationEntityFilters :: Lens' DescribeAffectedEntitiesForOrganization (NonEmpty EventAccountFilter)

A JSON set of elements including the awsAccountId and the eventArn.

Destructuring the Response

data DescribeAffectedEntitiesForOrganizationResponse

See: newDescribeAffectedEntitiesForOrganizationResponse smart constructor.

Constructors

DescribeAffectedEntitiesForOrganizationResponse'

Fields

entities :: Maybe [AffectedEntity]

A JSON set of elements including the awsAccountId and its entityArn, entityValue and its entityArn, lastUpdatedTime, and statusCode.

failedSet :: Maybe [OrganizationAffectedEntitiesErrorItem]

A JSON set of elements of the failed response, including the awsAccountId, errorMessage, errorName, and eventArn.

nextToken :: Maybe Text

If the results of a search are large, only a portion of the results are returned, and a nextToken pagination token is returned in the response. To retrieve the next batch of results, reissue the search request and include the returned token. When all results have been returned, the response does not contain a pagination token value.

httpStatus :: Int

The response's http status code.

newDescribeAffectedEntitiesForOrganizationResponse

Arguments

:: Int
$sel:httpStatus:DescribeAffectedEntitiesForOrganizationResponse'

-> DescribeAffectedEntitiesForOrganizationResponse

Create a value of DescribeAffectedEntitiesForOrganizationResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:entities:DescribeAffectedEntitiesForOrganizationResponse', describeAffectedEntitiesForOrganizationResponse_entities - A JSON set of elements including the awsAccountId and its entityArn, entityValue and its entityArn, lastUpdatedTime, and statusCode.

$sel:failedSet:DescribeAffectedEntitiesForOrganizationResponse', describeAffectedEntitiesForOrganizationResponse_failedSet - A JSON set of elements of the failed response, including the awsAccountId, errorMessage, errorName, and eventArn.

$sel:nextToken:DescribeAffectedEntitiesForOrganization', describeAffectedEntitiesForOrganizationResponse_nextToken - If the results of a search are large, only a portion of the results are returned, and a nextToken pagination token is returned in the response. To retrieve the next batch of results, reissue the search request and include the returned token. When all results have been returned, the response does not contain a pagination token value.

$sel:httpStatus:DescribeAffectedEntitiesForOrganizationResponse', describeAffectedEntitiesForOrganizationResponse_httpStatus - The response's http status code.

Response Lenses

describeAffectedEntitiesForOrganizationResponse_entities :: Lens' DescribeAffectedEntitiesForOrganizationResponse (Maybe [AffectedEntity])

A JSON set of elements including the awsAccountId and its entityArn, entityValue and its entityArn, lastUpdatedTime, and statusCode.

describeAffectedEntitiesForOrganizationResponse_failedSet :: Lens' DescribeAffectedEntitiesForOrganizationResponse (Maybe [OrganizationAffectedEntitiesErrorItem])

A JSON set of elements of the failed response, including the awsAccountId, errorMessage, errorName, and eventArn.

describeAffectedEntitiesForOrganizationResponse_nextToken :: Lens' DescribeAffectedEntitiesForOrganizationResponse (Maybe Text)

If the results of a search are large, only a portion of the results are returned, and a nextToken pagination token is returned in the response. To retrieve the next batch of results, reissue the search request and include the returned token. When all results have been returned, the response does not contain a pagination token value.

describeAffectedEntitiesForOrganizationResponse_httpStatus :: Lens' DescribeAffectedEntitiesForOrganizationResponse Int

The response's http status code.