Amazonka.AWSHealth.DescribeEventDetailsForOrganization

Description

Returns detailed information about one or more specified events for one or more AWS accounts in your organization. This information includes standard event data (such as the AWS Region and service), an event description, and (depending on the event) possible metadata. This operation doesn't return affected entities, such as the resources related to the event. To return affected entities, use the DescribeAffectedEntitiesForOrganization operation.

Before you can call this operation, you must first enable AWS Health to work with AWS Organizations. To do this, call the EnableHealthServiceAccessForOrganization operation from your organization's management account.

When you call the DescribeEventDetailsForOrganization operation, specify the organizationEventDetailFilters object in the request. Depending on the AWS Health event type, note the following differences:

To return event details for a public event, you must specify a null value for the awsAccountId parameter. If you specify an account ID for a public event, AWS Health returns an error message because public events aren't specific to an account.

To return event details for an event that is specific to an account in your organization, you must specify the awsAccountId parameter in the request. If you don't specify an account ID, AWS Health returns an error message because the event is specific to an account in your organization.

For more information, see Event.

This operation doesn't support resource-level permissions. You can't use this operation to allow or deny access to specific AWS Health events. For more information, see Resource- and action-based conditions in the AWS Health User Guide.

Creating a Request

data DescribeEventDetailsForOrganization

See: newDescribeEventDetailsForOrganization smart constructor.

Constructors

DescribeEventDetailsForOrganization'

Fields

locale :: Maybe Text

The locale (language) to return information in. English (en) is the default and the only supported value at this time.

organizationEventDetailFilters :: NonEmpty EventAccountFilter

A set of JSON elements that includes the awsAccountId and the eventArn.

newDescribeEventDetailsForOrganization

Arguments

:: NonEmpty EventAccountFilter
$sel:organizationEventDetailFilters:DescribeEventDetailsForOrganization'

-> DescribeEventDetailsForOrganization

Create a value of DescribeEventDetailsForOrganization with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:locale:DescribeEventDetailsForOrganization', describeEventDetailsForOrganization_locale - The locale (language) to return information in. English (en) is the default and the only supported value at this time.

$sel:organizationEventDetailFilters:DescribeEventDetailsForOrganization', describeEventDetailsForOrganization_organizationEventDetailFilters - A set of JSON elements that includes the awsAccountId and the eventArn.

Request Lenses

describeEventDetailsForOrganization_locale :: Lens' DescribeEventDetailsForOrganization (Maybe Text)

The locale (language) to return information in. English (en) is the default and the only supported value at this time.

describeEventDetailsForOrganization_organizationEventDetailFilters :: Lens' DescribeEventDetailsForOrganization (NonEmpty EventAccountFilter)

A set of JSON elements that includes the awsAccountId and the eventArn.

Destructuring the Response

data DescribeEventDetailsForOrganizationResponse

See: newDescribeEventDetailsForOrganizationResponse smart constructor.

Constructors

DescribeEventDetailsForOrganizationResponse'

Fields

successfulSet :: Maybe [OrganizationEventDetails]

Information about the events that could be retrieved.

failedSet :: Maybe [OrganizationEventDetailsErrorItem]

Error messages for any events that could not be retrieved.

httpStatus :: Int

The response's http status code.

newDescribeEventDetailsForOrganizationResponse

Arguments

:: Int
$sel:httpStatus:DescribeEventDetailsForOrganizationResponse'

-> DescribeEventDetailsForOrganizationResponse

Create a value of DescribeEventDetailsForOrganizationResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:successfulSet:DescribeEventDetailsForOrganizationResponse', describeEventDetailsForOrganizationResponse_successfulSet - Information about the events that could be retrieved.

$sel:failedSet:DescribeEventDetailsForOrganizationResponse', describeEventDetailsForOrganizationResponse_failedSet - Error messages for any events that could not be retrieved.

$sel:httpStatus:DescribeEventDetailsForOrganizationResponse', describeEventDetailsForOrganizationResponse_httpStatus - The response's http status code.

Response Lenses

describeEventDetailsForOrganizationResponse_successfulSet :: Lens' DescribeEventDetailsForOrganizationResponse (Maybe [OrganizationEventDetails])

Information about the events that could be retrieved.

describeEventDetailsForOrganizationResponse_failedSet :: Lens' DescribeEventDetailsForOrganizationResponse (Maybe [OrganizationEventDetailsErrorItem])

Error messages for any events that could not be retrieved.

describeEventDetailsForOrganizationResponse_httpStatus :: Lens' DescribeEventDetailsForOrganizationResponse Int

The response's http status code.