Amazonka.AWSHealth.Types.Event

Description

Documentation

data Event

Summary information about an AWS Health event.

AWS Health events can be public or account-specific:

Public events might be service events that are not specific to an AWS account. For example, if there is an issue with an AWS Region, AWS Health provides information about the event, even if you don't use services or resources in that Region.

Account-specific events are specific to either your AWS account or an account in your organization. For example, if there's an issue with Amazon Elastic Compute Cloud in a Region that you use, AWS Health provides information about the event and the affected resources in the account.

You can determine if an event is public or account-specific by using the eventScopeCode parameter. For more information, see eventScopeCode.

See: newEvent smart constructor.

Constructors

Event'

Fields

lastUpdatedTime :: Maybe POSIX

The most recent date and time that the event was updated.

arn :: Maybe Text

The unique identifier for the event. The event ARN has the arn:aws:health:event-region::event/SERVICE/EVENT_TYPE_CODE/EVENT_TYPE_PLUS_ID format.

For example, an event ARN might look like the following:

arn:aws:health:us-east-1::event/EC2/EC2_INSTANCE_RETIREMENT_SCHEDULED/EC2_INSTANCE_RETIREMENT_SCHEDULED_ABC123-DEF456

service :: Maybe Text

The AWS service that is affected by the event. For example, EC2, RDS.

startTime :: Maybe POSIX

The date and time that the event began.

eventScopeCode :: Maybe EventScopeCode

This parameter specifies if the AWS Health event is a public AWS service event or an account-specific event.

If the eventScopeCode value is PUBLIC, then the affectedAccounts value is always empty.

If the eventScopeCode value is ACCOUNT_SPECIFIC, then the affectedAccounts value lists the affected AWS accounts in your organization. For example, if an event affects a service such as Amazon Elastic Compute Cloud and you have AWS accounts that use that service, those account IDs appear in the response.

If the eventScopeCode value is NONE, then the eventArn that you specified in the request is invalid or doesn't exist.

eventTypeCode :: Maybe Text

The unique identifier for the event type. The format is AWS_SERVICE_DESCRIPTION ; for example, AWS_EC2_SYSTEM_MAINTENANCE_EVENT.

eventTypeCategory :: Maybe EventTypeCategory

The category of the event. Possible values are issue, scheduledChange, and accountNotification.

availabilityZone :: Maybe Text

The AWS Availability Zone of the event. For example, us-east-1a.

endTime :: Maybe POSIX

The date and time that the event ended.

region :: Maybe Text

The AWS Region name of the event.

statusCode :: Maybe EventStatusCode

The most recent status of the event. Possible values are open, closed, and upcoming.

newEvent :: Event

Create a value of Event with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:lastUpdatedTime:Event', event_lastUpdatedTime - The most recent date and time that the event was updated.

$sel:arn:Event', event_arn - The unique identifier for the event. The event ARN has the arn:aws:health:event-region::event/SERVICE/EVENT_TYPE_CODE/EVENT_TYPE_PLUS_ID format.

For example, an event ARN might look like the following:

arn:aws:health:us-east-1::event/EC2/EC2_INSTANCE_RETIREMENT_SCHEDULED/EC2_INSTANCE_RETIREMENT_SCHEDULED_ABC123-DEF456

$sel:service:Event', event_service - The AWS service that is affected by the event. For example, EC2, RDS.

$sel:startTime:Event', event_startTime - The date and time that the event began.

$sel:eventScopeCode:Event', event_eventScopeCode - This parameter specifies if the AWS Health event is a public AWS service event or an account-specific event.

If the eventScopeCode value is PUBLIC, then the affectedAccounts value is always empty.

If the eventScopeCode value is ACCOUNT_SPECIFIC, then the affectedAccounts value lists the affected AWS accounts in your organization. For example, if an event affects a service such as Amazon Elastic Compute Cloud and you have AWS accounts that use that service, those account IDs appear in the response.

If the eventScopeCode value is NONE, then the eventArn that you specified in the request is invalid or doesn't exist.

$sel:eventTypeCode:Event', event_eventTypeCode - The unique identifier for the event type. The format is AWS_SERVICE_DESCRIPTION ; for example, AWS_EC2_SYSTEM_MAINTENANCE_EVENT.

$sel:eventTypeCategory:Event', event_eventTypeCategory - The category of the event. Possible values are issue, scheduledChange, and accountNotification.

$sel:availabilityZone:Event', event_availabilityZone - The AWS Availability Zone of the event. For example, us-east-1a.

$sel:endTime:Event', event_endTime - The date and time that the event ended.

$sel:region:Event', event_region - The AWS Region name of the event.

$sel:statusCode:Event', event_statusCode - The most recent status of the event. Possible values are open, closed, and upcoming.

event_lastUpdatedTime :: Lens' Event (Maybe UTCTime)

The most recent date and time that the event was updated.

event_arn :: Lens' Event (Maybe Text)

The unique identifier for the event. The event ARN has the arn:aws:health:event-region::event/SERVICE/EVENT_TYPE_CODE/EVENT_TYPE_PLUS_ID format.

For example, an event ARN might look like the following:

arn:aws:health:us-east-1::event/EC2/EC2_INSTANCE_RETIREMENT_SCHEDULED/EC2_INSTANCE_RETIREMENT_SCHEDULED_ABC123-DEF456

event_service :: Lens' Event (Maybe Text)

The AWS service that is affected by the event. For example, EC2, RDS.

event_startTime :: Lens' Event (Maybe UTCTime)

The date and time that the event began.

event_eventScopeCode :: Lens' Event (Maybe EventScopeCode)

This parameter specifies if the AWS Health event is a public AWS service event or an account-specific event.

If the eventScopeCode value is PUBLIC, then the affectedAccounts value is always empty.

If the eventScopeCode value is ACCOUNT_SPECIFIC, then the affectedAccounts value lists the affected AWS accounts in your organization. For example, if an event affects a service such as Amazon Elastic Compute Cloud and you have AWS accounts that use that service, those account IDs appear in the response.

If the eventScopeCode value is NONE, then the eventArn that you specified in the request is invalid or doesn't exist.

event_eventTypeCode :: Lens' Event (Maybe Text)

The unique identifier for the event type. The format is AWS_SERVICE_DESCRIPTION ; for example, AWS_EC2_SYSTEM_MAINTENANCE_EVENT.

event_eventTypeCategory :: Lens' Event (Maybe EventTypeCategory)

The category of the event. Possible values are issue, scheduledChange, and accountNotification.

event_availabilityZone :: Lens' Event (Maybe Text)

The AWS Availability Zone of the event. For example, us-east-1a.

event_endTime :: Lens' Event (Maybe UTCTime)

The date and time that the event ended.

event_region :: Lens' Event (Maybe Text)

The AWS Region name of the event.

event_statusCode :: Lens' Event (Maybe EventStatusCode)

The most recent status of the event. Possible values are open, closed, and upcoming.