Amazonka.AWSHealth.Types.EventTypeFilter

Description

Documentation

data EventTypeFilter

The values to use to filter results from the DescribeEventTypes operation.

See: newEventTypeFilter smart constructor.

Constructors

EventTypeFilter'

Fields

eventTypeCategories :: Maybe (NonEmpty EventTypeCategory)

A list of event type category codes (issue, scheduledChange, or accountNotification).

eventTypeCodes :: Maybe (NonEmpty Text)

A list of event type codes.

services :: Maybe (NonEmpty Text)

The AWS services associated with the event. For example, EC2, RDS.

newEventTypeFilter :: EventTypeFilter

Create a value of EventTypeFilter with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:eventTypeCategories:EventTypeFilter', eventTypeFilter_eventTypeCategories - A list of event type category codes (issue, scheduledChange, or accountNotification).

$sel:eventTypeCodes:EventTypeFilter', eventTypeFilter_eventTypeCodes - A list of event type codes.

$sel:services:EventTypeFilter', eventTypeFilter_services - The AWS services associated with the event. For example, EC2, RDS.

eventTypeFilter_eventTypeCategories :: Lens' EventTypeFilter (Maybe (NonEmpty EventTypeCategory))

A list of event type category codes (issue, scheduledChange, or accountNotification).

eventTypeFilter_eventTypeCodes :: Lens' EventTypeFilter (Maybe (NonEmpty Text))

A list of event type codes.

eventTypeFilter_services :: Lens' EventTypeFilter (Maybe (NonEmpty Text))

The AWS services associated with the event. For example, EC2, RDS.