Amazonka.AccessAnalyzer.GetAnalyzer

Description

Retrieves information about the specified analyzer.

Creating a Request

data GetAnalyzer

Retrieves an analyzer.

See: newGetAnalyzer smart constructor.

Constructors

GetAnalyzer'

Fields

analyzerName :: Text

The name of the analyzer retrieved.

newGetAnalyzer

Arguments

:: Text
$sel:analyzerName:GetAnalyzer'

-> GetAnalyzer

Create a value of GetAnalyzer with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:analyzerName:GetAnalyzer', getAnalyzer_analyzerName - The name of the analyzer retrieved.

Request Lenses

getAnalyzer_analyzerName :: Lens' GetAnalyzer Text

The name of the analyzer retrieved.

Destructuring the Response

data GetAnalyzerResponse

The response to the request.

See: newGetAnalyzerResponse smart constructor.

Constructors

GetAnalyzerResponse'

Fields

httpStatus :: Int

The response's http status code.

analyzer :: AnalyzerSummary

An AnalyzerSummary object that contains information about the analyzer.

newGetAnalyzerResponse

Arguments

:: Int
$sel:httpStatus:GetAnalyzerResponse'

-> AnalyzerSummary
$sel:analyzer:GetAnalyzerResponse'

-> GetAnalyzerResponse

Create a value of GetAnalyzerResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:httpStatus:GetAnalyzerResponse', getAnalyzerResponse_httpStatus - The response's http status code.

$sel:analyzer:GetAnalyzerResponse', getAnalyzerResponse_analyzer - An AnalyzerSummary object that contains information about the analyzer.

Response Lenses

getAnalyzerResponse_httpStatus :: Lens' GetAnalyzerResponse Int

The response's http status code.

getAnalyzerResponse_analyzer :: Lens' GetAnalyzerResponse AnalyzerSummary

An AnalyzerSummary object that contains information about the analyzer.