Amazonka.AccessAnalyzer.Lens

Description

Operations

ListFindings

listFindings_nextToken :: Lens' ListFindings (Maybe Text)

A token used for pagination of results returned.

listFindings_sort :: Lens' ListFindings (Maybe SortCriteria)

The sort order for the findings returned.

listFindings_filter :: Lens' ListFindings (Maybe (HashMap Text Criterion))

A filter to match for the findings to return.

listFindings_maxResults :: Lens' ListFindings (Maybe Int)

The maximum number of results to return in the response.

listFindings_analyzerArn :: Lens' ListFindings Text

The ARN of the analyzer to retrieve findings from.

listFindingsResponse_nextToken :: Lens' ListFindingsResponse (Maybe Text)

A token used for pagination of results returned.

listFindingsResponse_httpStatus :: Lens' ListFindingsResponse Int

The response's http status code.

listFindingsResponse_findings :: Lens' ListFindingsResponse [FindingSummary]

A list of findings retrieved from the analyzer that match the filter criteria specified, if any.

GetAnalyzedResource

getAnalyzedResource_analyzerArn :: Lens' GetAnalyzedResource Text

The ARN of the analyzer to retrieve information from.

getAnalyzedResource_resourceArn :: Lens' GetAnalyzedResource Text

The ARN of the resource to retrieve information about.

getAnalyzedResourceResponse_resource :: Lens' GetAnalyzedResourceResponse (Maybe AnalyzedResource)

An AnalyzedResource object that contains information that IAM Access Analyzer found when it analyzed the resource.

getAnalyzedResourceResponse_httpStatus :: Lens' GetAnalyzedResourceResponse Int

The response's http status code.

ListPolicyGenerations

listPolicyGenerations_nextToken :: Lens' ListPolicyGenerations (Maybe Text)

A token used for pagination of results returned.

listPolicyGenerations_principalArn :: Lens' ListPolicyGenerations (Maybe Text)

The ARN of the IAM entity (user or role) for which you are generating a policy. Use this with ListGeneratedPolicies to filter the results to only include results for a specific principal.

listPolicyGenerations_maxResults :: Lens' ListPolicyGenerations (Maybe Natural)

The maximum number of results to return in the response.

listPolicyGenerationsResponse_nextToken :: Lens' ListPolicyGenerationsResponse (Maybe Text)

A token used for pagination of results returned.

listPolicyGenerationsResponse_httpStatus :: Lens' ListPolicyGenerationsResponse Int

The response's http status code.

listPolicyGenerationsResponse_policyGenerations :: Lens' ListPolicyGenerationsResponse [PolicyGeneration]

A PolicyGeneration object that contains details about the generated policy.

ListAccessPreviews

listAccessPreviews_nextToken :: Lens' ListAccessPreviews (Maybe Text)

A token used for pagination of results returned.

listAccessPreviews_maxResults :: Lens' ListAccessPreviews (Maybe Int)

The maximum number of results to return in the response.

listAccessPreviews_analyzerArn :: Lens' ListAccessPreviews Text

The ARN of the analyzer used to generate the access preview.

listAccessPreviewsResponse_nextToken :: Lens' ListAccessPreviewsResponse (Maybe Text)

A token used for pagination of results returned.

listAccessPreviewsResponse_httpStatus :: Lens' ListAccessPreviewsResponse Int

The response's http status code.

listAccessPreviewsResponse_accessPreviews :: Lens' ListAccessPreviewsResponse [AccessPreviewSummary]

A list of access previews retrieved for the analyzer.

CreateAccessPreview

createAccessPreview_clientToken :: Lens' CreateAccessPreview (Maybe Text)

A client token.

createAccessPreview_analyzerArn :: Lens' CreateAccessPreview Text

The ARN of the account analyzer used to generate the access preview. You can only create an access preview for analyzers with an Account type and Active status.

createAccessPreview_configurations :: Lens' CreateAccessPreview (HashMap Text Configuration)

Access control configuration for your resource that is used to generate the access preview. The access preview includes findings for external access allowed to the resource with the proposed access control configuration. The configuration must contain exactly one element.

createAccessPreviewResponse_httpStatus :: Lens' CreateAccessPreviewResponse Int

The response's http status code.

createAccessPreviewResponse_id :: Lens' CreateAccessPreviewResponse Text

The unique ID for the access preview.

ListTagsForResource

listTagsForResource_resourceArn :: Lens' ListTagsForResource Text

The ARN of the resource to retrieve tags from.

listTagsForResourceResponse_tags :: Lens' ListTagsForResourceResponse (Maybe (HashMap Text Text))

The tags that are applied to the specified resource.

listTagsForResourceResponse_httpStatus :: Lens' ListTagsForResourceResponse Int

The response's http status code.

StartResourceScan

startResourceScan_analyzerArn :: Lens' StartResourceScan Text

The ARN of the analyzer to use to scan the policies applied to the specified resource.

startResourceScan_resourceArn :: Lens' StartResourceScan Text

The ARN of the resource to scan.

DeleteArchiveRule

deleteArchiveRule_clientToken :: Lens' DeleteArchiveRule (Maybe Text)

A client token.

deleteArchiveRule_analyzerName :: Lens' DeleteArchiveRule Text

The name of the analyzer that associated with the archive rule to delete.

deleteArchiveRule_ruleName :: Lens' DeleteArchiveRule Text

The name of the rule to delete.

UpdateArchiveRule

updateArchiveRule_clientToken :: Lens' UpdateArchiveRule (Maybe Text)

A client token.

updateArchiveRule_analyzerName :: Lens' UpdateArchiveRule Text

The name of the analyzer to update the archive rules for.

updateArchiveRule_filter :: Lens' UpdateArchiveRule (HashMap Text Criterion)

A filter to match for the rules to update. Only rules that match the filter are updated.

updateArchiveRule_ruleName :: Lens' UpdateArchiveRule Text

The name of the rule to update.

GetAccessPreview

getAccessPreview_accessPreviewId :: Lens' GetAccessPreview Text

The unique ID for the access preview.

getAccessPreview_analyzerArn :: Lens' GetAccessPreview Text

The ARN of the analyzer used to generate the access preview.

getAccessPreviewResponse_httpStatus :: Lens' GetAccessPreviewResponse Int

The response's http status code.

getAccessPreviewResponse_accessPreview :: Lens' GetAccessPreviewResponse AccessPreview

An object that contains information about the access preview.

ListAnalyzedResources

listAnalyzedResources_resourceType :: Lens' ListAnalyzedResources (Maybe ResourceType)

The type of resource.

listAnalyzedResources_nextToken :: Lens' ListAnalyzedResources (Maybe Text)

A token used for pagination of results returned.

listAnalyzedResources_maxResults :: Lens' ListAnalyzedResources (Maybe Int)

The maximum number of results to return in the response.

listAnalyzedResources_analyzerArn :: Lens' ListAnalyzedResources Text

The ARN of the analyzer to retrieve a list of analyzed resources from.

listAnalyzedResourcesResponse_nextToken :: Lens' ListAnalyzedResourcesResponse (Maybe Text)

A token used for pagination of results returned.

listAnalyzedResourcesResponse_httpStatus :: Lens' ListAnalyzedResourcesResponse Int

The response's http status code.

listAnalyzedResourcesResponse_analyzedResources :: Lens' ListAnalyzedResourcesResponse [AnalyzedResourceSummary]

A list of resources that were analyzed.

StartPolicyGeneration

startPolicyGeneration_clientToken :: Lens' StartPolicyGeneration (Maybe Text)

A unique, case-sensitive identifier that you provide to ensure the idempotency of the request. Idempotency ensures that an API request completes only once. With an idempotent request, if the original request completes successfully, the subsequent retries with the same client token return the result from the original successful request and they have no additional effect.

If you do not specify a client token, one is automatically generated by the Amazon Web Services SDK.

startPolicyGeneration_cloudTrailDetails :: Lens' StartPolicyGeneration (Maybe CloudTrailDetails)

A CloudTrailDetails object that contains details about a Trail that you want to analyze to generate policies.

startPolicyGeneration_policyGenerationDetails :: Lens' StartPolicyGeneration PolicyGenerationDetails

Contains the ARN of the IAM entity (user or role) for which you are generating a policy.

startPolicyGenerationResponse_httpStatus :: Lens' StartPolicyGenerationResponse Int

The response's http status code.

startPolicyGenerationResponse_jobId :: Lens' StartPolicyGenerationResponse Text

The JobId that is returned by the StartPolicyGeneration operation. The JobId can be used with GetGeneratedPolicy to retrieve the generated policies or used with CancelPolicyGeneration to cancel the policy generation request.

ValidatePolicy

validatePolicy_locale :: Lens' ValidatePolicy (Maybe Locale)

The locale to use for localizing the findings.

validatePolicy_nextToken :: Lens' ValidatePolicy (Maybe Text)

A token used for pagination of results returned.

validatePolicy_maxResults :: Lens' ValidatePolicy (Maybe Int)

The maximum number of results to return in the response.

validatePolicy_policyDocument :: Lens' ValidatePolicy Text

The JSON policy document to use as the content for the policy.

validatePolicy_policyType :: Lens' ValidatePolicy PolicyType

The type of policy to validate. Identity policies grant permissions to IAM principals. Identity policies include managed and inline policies for IAM roles, users, and groups. They also include service-control policies (SCPs) that are attached to an Amazon Web Services organization, organizational unit (OU), or an account.

Resource policies grant permissions on Amazon Web Services resources. Resource policies include trust policies for IAM roles and bucket policies for Amazon S3 buckets. You can provide a generic input such as identity policy or resource policy or a specific input such as managed policy or Amazon S3 bucket policy.

validatePolicyResponse_nextToken :: Lens' ValidatePolicyResponse (Maybe Text)

A token used for pagination of results returned.

validatePolicyResponse_httpStatus :: Lens' ValidatePolicyResponse Int

The response's http status code.

validatePolicyResponse_findings :: Lens' ValidatePolicyResponse [ValidatePolicyFinding]

The list of findings in a policy returned by IAM Access Analyzer based on its suite of policy checks.

DeleteAnalyzer

deleteAnalyzer_clientToken :: Lens' DeleteAnalyzer (Maybe Text)

A client token.

deleteAnalyzer_analyzerName :: Lens' DeleteAnalyzer Text

The name of the analyzer to delete.

UpdateFindings

updateFindings_clientToken :: Lens' UpdateFindings (Maybe Text)

A client token.

updateFindings_ids :: Lens' UpdateFindings (Maybe [Text])

The IDs of the findings to update.

updateFindings_resourceArn :: Lens' UpdateFindings (Maybe Text)

The ARN of the resource identified in the finding.

updateFindings_analyzerArn :: Lens' UpdateFindings Text

The ARN of the analyzer that generated the findings to update.

updateFindings_status :: Lens' UpdateFindings FindingStatusUpdate

The state represents the action to take to update the finding Status. Use ARCHIVE to change an Active finding to an Archived finding. Use ACTIVE to change an Archived finding to an Active finding.

ListAnalyzers

listAnalyzers_nextToken :: Lens' ListAnalyzers (Maybe Text)

A token used for pagination of results returned.

listAnalyzers_type :: Lens' ListAnalyzers (Maybe Type)

The type of analyzer.

listAnalyzers_maxResults :: Lens' ListAnalyzers (Maybe Int)

The maximum number of results to return in the response.

listAnalyzersResponse_nextToken :: Lens' ListAnalyzersResponse (Maybe Text)

A token used for pagination of results returned.

listAnalyzersResponse_httpStatus :: Lens' ListAnalyzersResponse Int

The response's http status code.

listAnalyzersResponse_analyzers :: Lens' ListAnalyzersResponse [AnalyzerSummary]

The analyzers retrieved.

ListAccessPreviewFindings

listAccessPreviewFindings_nextToken :: Lens' ListAccessPreviewFindings (Maybe Text)

A token used for pagination of results returned.

listAccessPreviewFindings_filter :: Lens' ListAccessPreviewFindings (Maybe (HashMap Text Criterion))

Criteria to filter the returned findings.

listAccessPreviewFindings_maxResults :: Lens' ListAccessPreviewFindings (Maybe Int)

The maximum number of results to return in the response.

listAccessPreviewFindings_accessPreviewId :: Lens' ListAccessPreviewFindings Text

The unique ID for the access preview.

listAccessPreviewFindings_analyzerArn :: Lens' ListAccessPreviewFindings Text

The ARN of the analyzer used to generate the access.

listAccessPreviewFindingsResponse_nextToken :: Lens' ListAccessPreviewFindingsResponse (Maybe Text)

A token used for pagination of results returned.

listAccessPreviewFindingsResponse_httpStatus :: Lens' ListAccessPreviewFindingsResponse Int

The response's http status code.

listAccessPreviewFindingsResponse_findings :: Lens' ListAccessPreviewFindingsResponse [AccessPreviewFinding]

A list of access preview findings that match the specified filter criteria.

GetArchiveRule

getArchiveRule_analyzerName :: Lens' GetArchiveRule Text

The name of the analyzer to retrieve rules from.

getArchiveRule_ruleName :: Lens' GetArchiveRule Text

The name of the rule to retrieve.

getArchiveRuleResponse_httpStatus :: Lens' GetArchiveRuleResponse Int

The response's http status code.

getArchiveRuleResponse_archiveRule :: Lens' GetArchiveRuleResponse ArchiveRuleSummary

Undocumented member.

CreateAnalyzer

createAnalyzer_clientToken :: Lens' CreateAnalyzer (Maybe Text)

A client token.

createAnalyzer_archiveRules :: Lens' CreateAnalyzer (Maybe [InlineArchiveRule])

Specifies the archive rules to add for the analyzer. Archive rules automatically archive findings that meet the criteria you define for the rule.

createAnalyzer_tags :: Lens' CreateAnalyzer (Maybe (HashMap Text Text))

The tags to apply to the analyzer.

createAnalyzer_analyzerName :: Lens' CreateAnalyzer Text

The name of the analyzer to create.

createAnalyzer_type :: Lens' CreateAnalyzer Type

The type of analyzer to create. Only ACCOUNT and ORGANIZATION analyzers are supported. You can create only one analyzer per account per Region. You can create up to 5 analyzers per organization per Region.

createAnalyzerResponse_arn :: Lens' CreateAnalyzerResponse (Maybe Text)

The ARN of the analyzer that was created by the request.

createAnalyzerResponse_httpStatus :: Lens' CreateAnalyzerResponse Int

The response's http status code.

ListArchiveRules

listArchiveRules_nextToken :: Lens' ListArchiveRules (Maybe Text)

A token used for pagination of results returned.

listArchiveRules_maxResults :: Lens' ListArchiveRules (Maybe Int)

The maximum number of results to return in the request.

listArchiveRules_analyzerName :: Lens' ListArchiveRules Text

The name of the analyzer to retrieve rules from.

listArchiveRulesResponse_nextToken :: Lens' ListArchiveRulesResponse (Maybe Text)

A token used for pagination of results returned.

listArchiveRulesResponse_httpStatus :: Lens' ListArchiveRulesResponse Int

The response's http status code.

listArchiveRulesResponse_archiveRules :: Lens' ListArchiveRulesResponse [ArchiveRuleSummary]

A list of archive rules created for the specified analyzer.

CreateArchiveRule

createArchiveRule_clientToken :: Lens' CreateArchiveRule (Maybe Text)

A client token.

createArchiveRule_analyzerName :: Lens' CreateArchiveRule Text

The name of the created analyzer.

createArchiveRule_filter :: Lens' CreateArchiveRule (HashMap Text Criterion)

The criteria for the rule.

createArchiveRule_ruleName :: Lens' CreateArchiveRule Text

The name of the rule to create.

CancelPolicyGeneration

cancelPolicyGeneration_jobId :: Lens' CancelPolicyGeneration Text

The JobId that is returned by the StartPolicyGeneration operation. The JobId can be used with GetGeneratedPolicy to retrieve the generated policies or used with CancelPolicyGeneration to cancel the policy generation request.

cancelPolicyGenerationResponse_httpStatus :: Lens' CancelPolicyGenerationResponse Int

The response's http status code.

TagResource

tagResource_resourceArn :: Lens' TagResource Text

The ARN of the resource to add the tag to.

tagResource_tags :: Lens' TagResource (HashMap Text Text)

The tags to add to the resource.

tagResourceResponse_httpStatus :: Lens' TagResourceResponse Int

The response's http status code.

ApplyArchiveRule

applyArchiveRule_clientToken :: Lens' ApplyArchiveRule (Maybe Text)

A client token.

applyArchiveRule_analyzerArn :: Lens' ApplyArchiveRule Text

The Amazon resource name (ARN) of the analyzer.

applyArchiveRule_ruleName :: Lens' ApplyArchiveRule Text

The name of the rule to apply.

GetAnalyzer

getAnalyzer_analyzerName :: Lens' GetAnalyzer Text

The name of the analyzer retrieved.

getAnalyzerResponse_httpStatus :: Lens' GetAnalyzerResponse Int

The response's http status code.

getAnalyzerResponse_analyzer :: Lens' GetAnalyzerResponse AnalyzerSummary

An AnalyzerSummary object that contains information about the analyzer.

UntagResource

untagResource_resourceArn :: Lens' UntagResource Text

The ARN of the resource to remove the tag from.

untagResource_tagKeys :: Lens' UntagResource [Text]

The key for the tag to add.

untagResourceResponse_httpStatus :: Lens' UntagResourceResponse Int

The response's http status code.

GetFinding

getFinding_analyzerArn :: Lens' GetFinding Text

The ARN of the analyzer that generated the finding.

getFinding_id :: Lens' GetFinding Text

The ID of the finding to retrieve.

getFindingResponse_finding :: Lens' GetFindingResponse (Maybe Finding)

A finding object that contains finding details.

getFindingResponse_httpStatus :: Lens' GetFindingResponse Int

The response's http status code.

GetGeneratedPolicy

getGeneratedPolicy_includeServiceLevelTemplate :: Lens' GetGeneratedPolicy (Maybe Bool)

The level of detail that you want to generate. You can specify whether to generate service-level policies.

IAM Access Analyzer uses iam:servicelastaccessed to identify services that have been used recently to create this service-level template.

getGeneratedPolicy_includeResourcePlaceholders :: Lens' GetGeneratedPolicy (Maybe Bool)

The level of detail that you want to generate. You can specify whether to generate policies with placeholders for resource ARNs for actions that support resource level granularity in policies.

For example, in the resource section of a policy, you can receive a placeholder such as "Resource":"arn:aws:s3:::${BucketName}" instead of "*".

getGeneratedPolicy_jobId :: Lens' GetGeneratedPolicy Text

The JobId that is returned by the StartPolicyGeneration operation. The JobId can be used with GetGeneratedPolicy to retrieve the generated policies or used with CancelPolicyGeneration to cancel the policy generation request.

getGeneratedPolicyResponse_httpStatus :: Lens' GetGeneratedPolicyResponse Int

The response's http status code.

getGeneratedPolicyResponse_generatedPolicyResult :: Lens' GetGeneratedPolicyResponse GeneratedPolicyResult

A GeneratedPolicyResult object that contains the generated policies and associated details.

getGeneratedPolicyResponse_jobDetails :: Lens' GetGeneratedPolicyResponse JobDetails

A GeneratedPolicyDetails object that contains details about the generated policy.

Types

AccessPreview

accessPreview_statusReason :: Lens' AccessPreview (Maybe AccessPreviewStatusReason)

Provides more details about the current status of the access preview.

For example, if the creation of the access preview fails, a Failed status is returned. This failure can be due to an internal issue with the analysis or due to an invalid resource configuration.

accessPreview_analyzerArn :: Lens' AccessPreview Text

The ARN of the analyzer used to generate the access preview.

accessPreview_configurations :: Lens' AccessPreview (HashMap Text Configuration)

A map of resource ARNs for the proposed resource configuration.

accessPreview_createdAt :: Lens' AccessPreview UTCTime

The time at which the access preview was created.

accessPreview_id :: Lens' AccessPreview Text

The unique ID for the access preview.

accessPreview_status :: Lens' AccessPreview AccessPreviewStatus

The status of the access preview.

Creating - The access preview creation is in progress.

Completed - The access preview is complete. You can preview findings for external access to the resource.

Failed - The access preview creation has failed.

AccessPreviewFinding

accessPreviewFinding_existingFindingStatus :: Lens' AccessPreviewFinding (Maybe FindingStatus)

The existing status of the finding, provided only for existing findings.

accessPreviewFinding_error :: Lens' AccessPreviewFinding (Maybe Text)

An error.

accessPreviewFinding_isPublic :: Lens' AccessPreviewFinding (Maybe Bool)

Indicates whether the policy that generated the finding allows public access to the resource.

accessPreviewFinding_action :: Lens' AccessPreviewFinding (Maybe [Text])

The action in the analyzed policy statement that an external principal has permission to perform.

accessPreviewFinding_sources :: Lens' AccessPreviewFinding (Maybe [FindingSource])

The sources of the finding. This indicates how the access that generated the finding is granted. It is populated for Amazon S3 bucket findings.

accessPreviewFinding_resource :: Lens' AccessPreviewFinding (Maybe Text)

The resource that an external principal has access to. This is the resource associated with the access preview.

accessPreviewFinding_principal :: Lens' AccessPreviewFinding (Maybe (HashMap Text Text))

The external principal that has access to a resource within the zone of trust.

accessPreviewFinding_existingFindingId :: Lens' AccessPreviewFinding (Maybe Text)

The existing ID of the finding in IAM Access Analyzer, provided only for existing findings.

accessPreviewFinding_condition :: Lens' AccessPreviewFinding (Maybe (HashMap Text Text))

The condition in the analyzed policy statement that resulted in a finding.

accessPreviewFinding_changeType :: Lens' AccessPreviewFinding FindingChangeType

Provides context on how the access preview finding compares to existing access identified in IAM Access Analyzer.

New - The finding is for newly-introduced access.

Unchanged - The preview finding is an existing finding that would remain unchanged.

Changed - The preview finding is an existing finding with a change in status.

For example, a Changed finding with preview status Resolved and existing status Active indicates the existing Active finding would become Resolved as a result of the proposed permissions change.

accessPreviewFinding_createdAt :: Lens' AccessPreviewFinding UTCTime

The time at which the access preview finding was created.

accessPreviewFinding_id :: Lens' AccessPreviewFinding Text

The ID of the access preview finding. This ID uniquely identifies the element in the list of access preview findings and is not related to the finding ID in Access Analyzer.

accessPreviewFinding_resourceOwnerAccount :: Lens' AccessPreviewFinding Text

The Amazon Web Services account ID that owns the resource. For most Amazon Web Services resources, the owning account is the account in which the resource was created.

accessPreviewFinding_resourceType :: Lens' AccessPreviewFinding ResourceType

The type of the resource that can be accessed in the finding.

accessPreviewFinding_status :: Lens' AccessPreviewFinding FindingStatus

The preview status of the finding. This is what the status of the finding would be after permissions deployment. For example, a Changed finding with preview status Resolved and existing status Active indicates the existing Active finding would become Resolved as a result of the proposed permissions change.

AccessPreviewStatusReason

accessPreviewStatusReason_code :: Lens' AccessPreviewStatusReason AccessPreviewStatusReasonCode

The reason code for the current status of the access preview.

AccessPreviewSummary

accessPreviewSummary_statusReason :: Lens' AccessPreviewSummary (Maybe AccessPreviewStatusReason)

Undocumented member.

accessPreviewSummary_analyzerArn :: Lens' AccessPreviewSummary Text

The ARN of the analyzer used to generate the access preview.

accessPreviewSummary_createdAt :: Lens' AccessPreviewSummary UTCTime

The time at which the access preview was created.

accessPreviewSummary_id :: Lens' AccessPreviewSummary Text

The unique ID for the access preview.

accessPreviewSummary_status :: Lens' AccessPreviewSummary AccessPreviewStatus

The status of the access preview.

Creating - The access preview creation is in progress.

Completed - The access preview is complete and previews the findings for external access to the resource.

Failed - The access preview creation has failed.

AclGrantee

aclGrantee_uri :: Lens' AclGrantee (Maybe Text)

Used for granting permissions to a predefined group.

aclGrantee_id :: Lens' AclGrantee (Maybe Text)

The value specified is the canonical user ID of an Amazon Web Services account.

AnalyzedResource

analyzedResource_status :: Lens' AnalyzedResource (Maybe FindingStatus)

The current status of the finding generated from the analyzed resource.

analyzedResource_actions :: Lens' AnalyzedResource (Maybe [Text])

The actions that an external principal is granted permission to use by the policy that generated the finding.

analyzedResource_error :: Lens' AnalyzedResource (Maybe Text)

An error message.

analyzedResource_sharedVia :: Lens' AnalyzedResource (Maybe [Text])

Indicates how the access that generated the finding is granted. This is populated for Amazon S3 bucket findings.

analyzedResource_analyzedAt :: Lens' AnalyzedResource UTCTime

The time at which the resource was analyzed.

analyzedResource_createdAt :: Lens' AnalyzedResource UTCTime

The time at which the finding was created.

analyzedResource_isPublic :: Lens' AnalyzedResource Bool

Indicates whether the policy that generated the finding grants public access to the resource.

analyzedResource_resourceArn :: Lens' AnalyzedResource Text

The ARN of the resource that was analyzed.

analyzedResource_resourceOwnerAccount :: Lens' AnalyzedResource Text

The Amazon Web Services account ID that owns the resource.

analyzedResource_resourceType :: Lens' AnalyzedResource ResourceType

The type of the resource that was analyzed.

analyzedResource_updatedAt :: Lens' AnalyzedResource UTCTime

The time at which the finding was updated.

AnalyzedResourceSummary

analyzedResourceSummary_resourceArn :: Lens' AnalyzedResourceSummary Text

The ARN of the analyzed resource.

analyzedResourceSummary_resourceOwnerAccount :: Lens' AnalyzedResourceSummary Text

The Amazon Web Services account ID that owns the resource.

analyzedResourceSummary_resourceType :: Lens' AnalyzedResourceSummary ResourceType

The type of resource that was analyzed.

AnalyzerSummary

analyzerSummary_lastResourceAnalyzedAt :: Lens' AnalyzerSummary (Maybe UTCTime)

The time at which the most recently analyzed resource was analyzed.

analyzerSummary_lastResourceAnalyzed :: Lens' AnalyzerSummary (Maybe Text)

The resource that was most recently analyzed by the analyzer.

analyzerSummary_statusReason :: Lens' AnalyzerSummary (Maybe StatusReason)

The statusReason provides more details about the current status of the analyzer. For example, if the creation for the analyzer fails, a Failed status is returned. For an analyzer with organization as the type, this failure can be due to an issue with creating the service-linked roles required in the member accounts of the Amazon Web Services organization.

analyzerSummary_tags :: Lens' AnalyzerSummary (Maybe (HashMap Text Text))

The tags added to the analyzer.

analyzerSummary_arn :: Lens' AnalyzerSummary Text

The ARN of the analyzer.

analyzerSummary_createdAt :: Lens' AnalyzerSummary UTCTime

A timestamp for the time at which the analyzer was created.

analyzerSummary_name :: Lens' AnalyzerSummary Text

The name of the analyzer.

analyzerSummary_status :: Lens' AnalyzerSummary AnalyzerStatus

The status of the analyzer. An Active analyzer successfully monitors supported resources and generates new findings. The analyzer is Disabled when a user action, such as removing trusted access for Identity and Access Management Access Analyzer from Organizations, causes the analyzer to stop generating new findings. The status is Creating when the analyzer creation is in progress and Failed when the analyzer creation has failed.

analyzerSummary_type :: Lens' AnalyzerSummary Type

The type of analyzer, which corresponds to the zone of trust chosen for the analyzer.

ArchiveRuleSummary

archiveRuleSummary_createdAt :: Lens' ArchiveRuleSummary UTCTime

The time at which the archive rule was created.

archiveRuleSummary_filter :: Lens' ArchiveRuleSummary (HashMap Text Criterion)

A filter used to define the archive rule.

archiveRuleSummary_ruleName :: Lens' ArchiveRuleSummary Text

The name of the archive rule.

archiveRuleSummary_updatedAt :: Lens' ArchiveRuleSummary UTCTime

The time at which the archive rule was last updated.

CloudTrailDetails

cloudTrailDetails_endTime :: Lens' CloudTrailDetails (Maybe UTCTime)

The end of the time range for which IAM Access Analyzer reviews your CloudTrail events. Events with a timestamp after this time are not considered to generate a policy. If this is not included in the request, the default value is the current time.

cloudTrailDetails_accessRole :: Lens' CloudTrailDetails Text

The ARN of the service role that IAM Access Analyzer uses to access your CloudTrail trail and service last accessed information.

cloudTrailDetails_startTime :: Lens' CloudTrailDetails UTCTime

The start of the time range for which IAM Access Analyzer reviews your CloudTrail events. Events with a timestamp before this time are not considered to generate a policy.

cloudTrailDetails_trails :: Lens' CloudTrailDetails [Trail]

A Trail object that contains settings for a trail.

CloudTrailProperties

cloudTrailProperties_endTime :: Lens' CloudTrailProperties UTCTime

The end of the time range for which IAM Access Analyzer reviews your CloudTrail events. Events with a timestamp after this time are not considered to generate a policy. If this is not included in the request, the default value is the current time.

cloudTrailProperties_startTime :: Lens' CloudTrailProperties UTCTime

The start of the time range for which IAM Access Analyzer reviews your CloudTrail events. Events with a timestamp before this time are not considered to generate a policy.

cloudTrailProperties_trailProperties :: Lens' CloudTrailProperties [TrailProperties]

A TrailProperties object that contains settings for trail properties.

Configuration

configuration_kmsKey :: Lens' Configuration (Maybe KmsKeyConfiguration)

The access control configuration is for a KMS key.

configuration_secretsManagerSecret :: Lens' Configuration (Maybe SecretsManagerSecretConfiguration)

The access control configuration is for a Secrets Manager secret.

configuration_sqsQueue :: Lens' Configuration (Maybe SqsQueueConfiguration)

The access control configuration is for an Amazon SQS queue.

configuration_s3Bucket :: Lens' Configuration (Maybe S3BucketConfiguration)

The access control configuration is for an Amazon S3 Bucket.

configuration_iamRole :: Lens' Configuration (Maybe IamRoleConfiguration)

The access control configuration is for an IAM role.

Criterion

criterion_eq :: Lens' Criterion (Maybe (NonEmpty Text))

An "equals" operator to match for the filter used to create the rule.

criterion_exists :: Lens' Criterion (Maybe Bool)

An "exists" operator to match for the filter used to create the rule.

criterion_neq :: Lens' Criterion (Maybe (NonEmpty Text))

A "not equals" operator to match for the filter used to create the rule.

criterion_contains :: Lens' Criterion (Maybe (NonEmpty Text))

A "contains" operator to match for the filter used to create the rule.

Finding

finding_error :: Lens' Finding (Maybe Text)

An error.

finding_isPublic :: Lens' Finding (Maybe Bool)

Indicates whether the policy that generated the finding allows public access to the resource.

finding_action :: Lens' Finding (Maybe [Text])

The action in the analyzed policy statement that an external principal has permission to use.

finding_sources :: Lens' Finding (Maybe [FindingSource])

The sources of the finding. This indicates how the access that generated the finding is granted. It is populated for Amazon S3 bucket findings.

finding_resource :: Lens' Finding (Maybe Text)

The resource that an external principal has access to.

finding_principal :: Lens' Finding (Maybe (HashMap Text Text))

The external principal that access to a resource within the zone of trust.

finding_analyzedAt :: Lens' Finding UTCTime

The time at which the resource was analyzed.

finding_condition :: Lens' Finding (HashMap Text Text)

The condition in the analyzed policy statement that resulted in a finding.

finding_createdAt :: Lens' Finding UTCTime

The time at which the finding was generated.

finding_id :: Lens' Finding Text

The ID of the finding.

finding_resourceOwnerAccount :: Lens' Finding Text

The Amazon Web Services account ID that owns the resource.

finding_resourceType :: Lens' Finding ResourceType

The type of the resource identified in the finding.

finding_status :: Lens' Finding FindingStatus

The current status of the finding.

finding_updatedAt :: Lens' Finding UTCTime

The time at which the finding was updated.

FindingSource

findingSource_detail :: Lens' FindingSource (Maybe FindingSourceDetail)

Includes details about how the access that generated the finding is granted. This is populated for Amazon S3 bucket findings.

findingSource_type :: Lens' FindingSource FindingSourceType

Indicates the type of access that generated the finding.

FindingSourceDetail

findingSourceDetail_accessPointArn :: Lens' FindingSourceDetail (Maybe Text)

The ARN of the access point that generated the finding. The ARN format depends on whether the ARN represents an access point or a multi-region access point.

FindingSummary

findingSummary_error :: Lens' FindingSummary (Maybe Text)

The error that resulted in an Error finding.

findingSummary_isPublic :: Lens' FindingSummary (Maybe Bool)

Indicates whether the finding reports a resource that has a policy that allows public access.

findingSummary_action :: Lens' FindingSummary (Maybe [Text])

The action in the analyzed policy statement that an external principal has permission to use.

findingSummary_sources :: Lens' FindingSummary (Maybe [FindingSource])

The sources of the finding. This indicates how the access that generated the finding is granted. It is populated for Amazon S3 bucket findings.

findingSummary_resource :: Lens' FindingSummary (Maybe Text)

The resource that the external principal has access to.

findingSummary_principal :: Lens' FindingSummary (Maybe (HashMap Text Text))

The external principal that has access to a resource within the zone of trust.

findingSummary_analyzedAt :: Lens' FindingSummary UTCTime

The time at which the resource-based policy that generated the finding was analyzed.

findingSummary_condition :: Lens' FindingSummary (HashMap Text Text)

The condition in the analyzed policy statement that resulted in a finding.

findingSummary_createdAt :: Lens' FindingSummary UTCTime

The time at which the finding was created.

findingSummary_id :: Lens' FindingSummary Text

The ID of the finding.

findingSummary_resourceOwnerAccount :: Lens' FindingSummary Text

The Amazon Web Services account ID that owns the resource.

findingSummary_resourceType :: Lens' FindingSummary ResourceType

The type of the resource that the external principal has access to.

findingSummary_status :: Lens' FindingSummary FindingStatus

The status of the finding.

findingSummary_updatedAt :: Lens' FindingSummary UTCTime

The time at which the finding was most recently updated.

GeneratedPolicy

generatedPolicy_policy :: Lens' GeneratedPolicy Text

The text to use as the content for the new policy. The policy is created using the CreatePolicy action.

GeneratedPolicyProperties

generatedPolicyProperties_isComplete :: Lens' GeneratedPolicyProperties (Maybe Bool)

This value is set to true if the generated policy contains all possible actions for a service that IAM Access Analyzer identified from the CloudTrail trail that you specified, and false otherwise.

generatedPolicyProperties_cloudTrailProperties :: Lens' GeneratedPolicyProperties (Maybe CloudTrailProperties)

Lists details about the Trail used to generated policy.

generatedPolicyProperties_principalArn :: Lens' GeneratedPolicyProperties Text

The ARN of the IAM entity (user or role) for which you are generating a policy.

GeneratedPolicyResult

generatedPolicyResult_generatedPolicies :: Lens' GeneratedPolicyResult (Maybe [GeneratedPolicy])

The text to use as the content for the new policy. The policy is created using the CreatePolicy action.

generatedPolicyResult_properties :: Lens' GeneratedPolicyResult GeneratedPolicyProperties

A GeneratedPolicyProperties object that contains properties of the generated policy.

IamRoleConfiguration

iamRoleConfiguration_trustPolicy :: Lens' IamRoleConfiguration (Maybe Text)

The proposed trust policy for the IAM role.

InlineArchiveRule

inlineArchiveRule_filter :: Lens' InlineArchiveRule (HashMap Text Criterion)

The condition and values for a criterion.

inlineArchiveRule_ruleName :: Lens' InlineArchiveRule Text

The name of the rule.

InternetConfiguration

JobDetails

jobDetails_completedOn :: Lens' JobDetails (Maybe UTCTime)

A timestamp of when the job was completed.

jobDetails_jobError :: Lens' JobDetails (Maybe JobError)

The job error for the policy generation request.

jobDetails_jobId :: Lens' JobDetails Text

The JobId that is returned by the StartPolicyGeneration operation. The JobId can be used with GetGeneratedPolicy to retrieve the generated policies or used with CancelPolicyGeneration to cancel the policy generation request.

jobDetails_startedOn :: Lens' JobDetails UTCTime

A timestamp of when the job was started.

jobDetails_status :: Lens' JobDetails JobStatus

The status of the job request.

JobError

jobError_code :: Lens' JobError JobErrorCode

The job error code.

jobError_message :: Lens' JobError Text

Specific information about the error. For example, which service quota was exceeded or which resource was not found.

KmsGrantConfiguration

kmsGrantConfiguration_retiringPrincipal :: Lens' KmsGrantConfiguration (Maybe Text)

The principal that is given permission to retire the grant by using RetireGrant operation.

kmsGrantConfiguration_constraints :: Lens' KmsGrantConfiguration (Maybe KmsGrantConstraints)

Use this structure to propose allowing cryptographic operations in the grant only when the operation request includes the specified encryption context.

kmsGrantConfiguration_granteePrincipal :: Lens' KmsGrantConfiguration Text

The principal that is given permission to perform the operations that the grant permits.

kmsGrantConfiguration_issuingAccount :: Lens' KmsGrantConfiguration Text

The Amazon Web Services account under which the grant was issued. The account is used to propose KMS grants issued by accounts other than the owner of the key.

kmsGrantConfiguration_operations :: Lens' KmsGrantConfiguration [KmsGrantOperation]

A list of operations that the grant permits.

KmsGrantConstraints

kmsGrantConstraints_encryptionContextEquals :: Lens' KmsGrantConstraints (Maybe (HashMap Text Text))

A list of key-value pairs that must match the encryption context in the cryptographic operation request. The grant allows the operation only when the encryption context in the request is the same as the encryption context specified in this constraint.

kmsGrantConstraints_encryptionContextSubset :: Lens' KmsGrantConstraints (Maybe (HashMap Text Text))

A list of key-value pairs that must be included in the encryption context of the cryptographic operation request. The grant allows the cryptographic operation only when the encryption context in the request includes the key-value pairs specified in this constraint, although it can include additional key-value pairs.

KmsKeyConfiguration

kmsKeyConfiguration_grants :: Lens' KmsKeyConfiguration (Maybe [KmsGrantConfiguration])

A list of proposed grant configurations for the KMS key. If the proposed grant configuration is for an existing key, the access preview uses the proposed list of grant configurations in place of the existing grants. Otherwise, the access preview uses the existing grants for the key.

kmsKeyConfiguration_keyPolicies :: Lens' KmsKeyConfiguration (Maybe (HashMap Text Text))

Resource policy configuration for the KMS key. The only valid value for the name of the key policy is default. For more information, see Default key policy.

Location

location_path :: Lens' Location [PathElement]

A path in a policy, represented as a sequence of path elements.

location_span :: Lens' Location Span

A span in a policy.

NetworkOriginConfiguration

networkOriginConfiguration_internetConfiguration :: Lens' NetworkOriginConfiguration (Maybe InternetConfiguration)

The configuration for the Amazon S3 access point or multi-region access point with an Internet origin.

networkOriginConfiguration_vpcConfiguration :: Lens' NetworkOriginConfiguration (Maybe VpcConfiguration)

Undocumented member.

PathElement

pathElement_value :: Lens' PathElement (Maybe Text)

Refers to the value associated with a given key in a JSON object.

pathElement_substring :: Lens' PathElement (Maybe Substring)

Refers to a substring of a literal string in a JSON object.

pathElement_key :: Lens' PathElement (Maybe Text)

Refers to a key in a JSON object.

pathElement_index :: Lens' PathElement (Maybe Int)

Refers to an index in a JSON array.

PolicyGeneration

policyGeneration_completedOn :: Lens' PolicyGeneration (Maybe UTCTime)

A timestamp of when the policy generation was completed.

policyGeneration_jobId :: Lens' PolicyGeneration Text

The JobId that is returned by the StartPolicyGeneration operation. The JobId can be used with GetGeneratedPolicy to retrieve the generated policies or used with CancelPolicyGeneration to cancel the policy generation request.

policyGeneration_principalArn :: Lens' PolicyGeneration Text

The ARN of the IAM entity (user or role) for which you are generating a policy.

policyGeneration_startedOn :: Lens' PolicyGeneration UTCTime

A timestamp of when the policy generation started.

policyGeneration_status :: Lens' PolicyGeneration JobStatus

The status of the policy generation request.

PolicyGenerationDetails

policyGenerationDetails_principalArn :: Lens' PolicyGenerationDetails Text

The ARN of the IAM entity (user or role) for which you are generating a policy.

Position

position_column :: Lens' Position Int

The column of the position, starting from 0.

position_line :: Lens' Position Int

The line of the position, starting from 1.

position_offset :: Lens' Position Int

The offset within the policy that corresponds to the position, starting from 0.

S3AccessPointConfiguration

s3AccessPointConfiguration_publicAccessBlock :: Lens' S3AccessPointConfiguration (Maybe S3PublicAccessBlockConfiguration)

The proposed S3PublicAccessBlock configuration to apply to this Amazon S3 access point or multi-region access point.

s3AccessPointConfiguration_accessPointPolicy :: Lens' S3AccessPointConfiguration (Maybe Text)

The access point or multi-region access point policy.

s3AccessPointConfiguration_networkOrigin :: Lens' S3AccessPointConfiguration (Maybe NetworkOriginConfiguration)

The proposed Internet and VpcConfiguration to apply to this Amazon S3 access point. VpcConfiguration does not apply to multi-region access points. If the access preview is for a new resource and neither is specified, the access preview uses Internet for the network origin. If the access preview is for an existing resource and neither is specified, the access preview uses the exiting network origin.

S3BucketAclGrantConfiguration

s3BucketAclGrantConfiguration_grantee :: Lens' S3BucketAclGrantConfiguration AclGrantee

The grantee to whom you’re assigning access rights.

s3BucketAclGrantConfiguration_permission :: Lens' S3BucketAclGrantConfiguration AclPermission

The permissions being granted.

S3BucketConfiguration

s3BucketConfiguration_accessPoints :: Lens' S3BucketConfiguration (Maybe (HashMap Text S3AccessPointConfiguration))

The configuration of Amazon S3 access points or multi-region access points for the bucket. You can propose up to 10 new access points per bucket.

s3BucketConfiguration_bucketPublicAccessBlock :: Lens' S3BucketConfiguration (Maybe S3PublicAccessBlockConfiguration)

The proposed block public access configuration for the Amazon S3 bucket.

s3BucketConfiguration_bucketAclGrants :: Lens' S3BucketConfiguration (Maybe [S3BucketAclGrantConfiguration])

The proposed list of ACL grants for the Amazon S3 bucket. You can propose up to 100 ACL grants per bucket. If the proposed grant configuration is for an existing bucket, the access preview uses the proposed list of grant configurations in place of the existing grants. Otherwise, the access preview uses the existing grants for the bucket.

s3BucketConfiguration_bucketPolicy :: Lens' S3BucketConfiguration (Maybe Text)

The proposed bucket policy for the Amazon S3 bucket.

S3PublicAccessBlockConfiguration

s3PublicAccessBlockConfiguration_ignorePublicAcls :: Lens' S3PublicAccessBlockConfiguration Bool

Specifies whether Amazon S3 should ignore public ACLs for this bucket and objects in this bucket.

s3PublicAccessBlockConfiguration_restrictPublicBuckets :: Lens' S3PublicAccessBlockConfiguration Bool

Specifies whether Amazon S3 should restrict public bucket policies for this bucket.

SecretsManagerSecretConfiguration

secretsManagerSecretConfiguration_kmsKeyId :: Lens' SecretsManagerSecretConfiguration (Maybe Text)

The proposed ARN, key ID, or alias of the KMS customer master key (CMK).

secretsManagerSecretConfiguration_secretPolicy :: Lens' SecretsManagerSecretConfiguration (Maybe Text)

The proposed resource policy defining who can access or manage the secret.

SortCriteria

sortCriteria_orderBy :: Lens' SortCriteria (Maybe OrderBy)

The sort order, ascending or descending.

sortCriteria_attributeName :: Lens' SortCriteria (Maybe Text)

The name of the attribute to sort on.

Span

span_end :: Lens' Span Position

The end position of the span (exclusive).

span_start :: Lens' Span Position

The start position of the span (inclusive).

SqsQueueConfiguration

sqsQueueConfiguration_queuePolicy :: Lens' SqsQueueConfiguration (Maybe Text)

The proposed resource policy for the Amazon SQS queue.

StatusReason

statusReason_code :: Lens' StatusReason ReasonCode

The reason code for the current status of the analyzer.

Substring

substring_length :: Lens' Substring Int

The length of the substring.

substring_start :: Lens' Substring Int

The start index of the substring, starting from 0.

Trail

trail_regions :: Lens' Trail (Maybe [Text])

A list of regions to get CloudTrail data from and analyze to generate a policy.

trail_allRegions :: Lens' Trail (Maybe Bool)

Possible values are true or false. If set to true, IAM Access Analyzer retrieves CloudTrail data from all regions to analyze and generate a policy.

trail_cloudTrailArn :: Lens' Trail Text

Specifies the ARN of the trail. The format of a trail ARN is arn:aws:cloudtrail:us-east-2:123456789012:trail/MyTrail.

TrailProperties

trailProperties_regions :: Lens' TrailProperties (Maybe [Text])

A list of regions to get CloudTrail data from and analyze to generate a policy.

trailProperties_allRegions :: Lens' TrailProperties (Maybe Bool)

Possible values are true or false. If set to true, IAM Access Analyzer retrieves CloudTrail data from all regions to analyze and generate a policy.

trailProperties_cloudTrailArn :: Lens' TrailProperties Text

Specifies the ARN of the trail. The format of a trail ARN is arn:aws:cloudtrail:us-east-2:123456789012:trail/MyTrail.

ValidatePolicyFinding

validatePolicyFinding_findingDetails :: Lens' ValidatePolicyFinding Text

A localized message that explains the finding and provides guidance on how to address it.

validatePolicyFinding_findingType :: Lens' ValidatePolicyFinding ValidatePolicyFindingType

The impact of the finding.

Security warnings report when the policy allows access that we consider overly permissive.

Errors report when a part of the policy is not functional.

Warnings report non-security issues when a policy does not conform to policy writing best practices.

Suggestions recommend stylistic improvements in the policy that do not impact access.

validatePolicyFinding_issueCode :: Lens' ValidatePolicyFinding Text

The issue code provides an identifier of the issue associated with this finding.

validatePolicyFinding_learnMoreLink :: Lens' ValidatePolicyFinding Text

A link to additional documentation about the type of finding.

validatePolicyFinding_locations :: Lens' ValidatePolicyFinding [Location]

The list of locations in the policy document that are related to the finding. The issue code provides a summary of an issue identified by the finding.

VpcConfiguration

vpcConfiguration_vpcId :: Lens' VpcConfiguration Text

If this field is specified, this access point will only allow connections from the specified VPC ID.