Amazonka.AccessAnalyzer.ListAccessPreviews

Description

Retrieves a list of access previews for the specified analyzer.

This operation returns paginated results.

Creating a Request

data ListAccessPreviews

See: newListAccessPreviews smart constructor.

Constructors

ListAccessPreviews'

Fields

nextToken :: Maybe Text

A token used for pagination of results returned.

maxResults :: Maybe Int

The maximum number of results to return in the response.

analyzerArn :: Text

The ARN of the analyzer used to generate the access preview.

newListAccessPreviews

Arguments

:: Text
$sel:analyzerArn:ListAccessPreviews'

-> ListAccessPreviews

Create a value of ListAccessPreviews with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:nextToken:ListAccessPreviews', listAccessPreviews_nextToken - A token used for pagination of results returned.

$sel:maxResults:ListAccessPreviews', listAccessPreviews_maxResults - The maximum number of results to return in the response.

$sel:analyzerArn:ListAccessPreviews', listAccessPreviews_analyzerArn - The ARN of the analyzer used to generate the access preview.

Request Lenses

listAccessPreviews_nextToken :: Lens' ListAccessPreviews (Maybe Text)

A token used for pagination of results returned.

listAccessPreviews_maxResults :: Lens' ListAccessPreviews (Maybe Int)

The maximum number of results to return in the response.

listAccessPreviews_analyzerArn :: Lens' ListAccessPreviews Text

The ARN of the analyzer used to generate the access preview.

Destructuring the Response

data ListAccessPreviewsResponse

See: newListAccessPreviewsResponse smart constructor.

Constructors

ListAccessPreviewsResponse'

Fields

nextToken :: Maybe Text

A token used for pagination of results returned.

httpStatus :: Int

The response's http status code.

accessPreviews :: [AccessPreviewSummary]

A list of access previews retrieved for the analyzer.

newListAccessPreviewsResponse

Arguments

:: Int
$sel:httpStatus:ListAccessPreviewsResponse'

-> ListAccessPreviewsResponse

Create a value of ListAccessPreviewsResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:nextToken:ListAccessPreviews', listAccessPreviewsResponse_nextToken - A token used for pagination of results returned.

$sel:httpStatus:ListAccessPreviewsResponse', listAccessPreviewsResponse_httpStatus - The response's http status code.

$sel:accessPreviews:ListAccessPreviewsResponse', listAccessPreviewsResponse_accessPreviews - A list of access previews retrieved for the analyzer.

Response Lenses

listAccessPreviewsResponse_nextToken :: Lens' ListAccessPreviewsResponse (Maybe Text)

A token used for pagination of results returned.

listAccessPreviewsResponse_httpStatus :: Lens' ListAccessPreviewsResponse Int

The response's http status code.

listAccessPreviewsResponse_accessPreviews :: Lens' ListAccessPreviewsResponse [AccessPreviewSummary]

A list of access previews retrieved for the analyzer.