Amazonka.AccessAnalyzer.ListFindings

Description

Retrieves a list of findings generated by the specified analyzer.

To learn about filter keys that you can use to retrieve a list of findings, see IAM Access Analyzer filter keys in the IAM User Guide.

This operation returns paginated results.

Creating a Request

data ListFindings

Retrieves a list of findings generated by the specified analyzer.

See: newListFindings smart constructor.

Constructors

ListFindings'

Fields

nextToken :: Maybe Text

A token used for pagination of results returned.

sort :: Maybe SortCriteria

The sort order for the findings returned.

filter' :: Maybe (HashMap Text Criterion)

A filter to match for the findings to return.

maxResults :: Maybe Int

The maximum number of results to return in the response.

analyzerArn :: Text

The ARN of the analyzer to retrieve findings from.

newListFindings

Arguments

:: Text
$sel:analyzerArn:ListFindings'

-> ListFindings

Create a value of ListFindings with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:nextToken:ListFindings', listFindings_nextToken - A token used for pagination of results returned.

$sel:sort:ListFindings', listFindings_sort - The sort order for the findings returned.

$sel:filter':ListFindings', listFindings_filter - A filter to match for the findings to return.

$sel:maxResults:ListFindings', listFindings_maxResults - The maximum number of results to return in the response.

$sel:analyzerArn:ListFindings', listFindings_analyzerArn - The ARN of the analyzer to retrieve findings from.

Request Lenses

listFindings_nextToken :: Lens' ListFindings (Maybe Text)

A token used for pagination of results returned.

listFindings_sort :: Lens' ListFindings (Maybe SortCriteria)

The sort order for the findings returned.

listFindings_filter :: Lens' ListFindings (Maybe (HashMap Text Criterion))

A filter to match for the findings to return.

listFindings_maxResults :: Lens' ListFindings (Maybe Int)

The maximum number of results to return in the response.

listFindings_analyzerArn :: Lens' ListFindings Text

The ARN of the analyzer to retrieve findings from.

Destructuring the Response

data ListFindingsResponse

The response to the request.

See: newListFindingsResponse smart constructor.

Constructors

ListFindingsResponse'

Fields

nextToken :: Maybe Text

A token used for pagination of results returned.

httpStatus :: Int

The response's http status code.

findings :: [FindingSummary]

A list of findings retrieved from the analyzer that match the filter criteria specified, if any.

newListFindingsResponse

Arguments

:: Int
$sel:httpStatus:ListFindingsResponse'

-> ListFindingsResponse

Create a value of ListFindingsResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:nextToken:ListFindings', listFindingsResponse_nextToken - A token used for pagination of results returned.

$sel:httpStatus:ListFindingsResponse', listFindingsResponse_httpStatus - The response's http status code.

$sel:findings:ListFindingsResponse', listFindingsResponse_findings - A list of findings retrieved from the analyzer that match the filter criteria specified, if any.

Response Lenses

listFindingsResponse_nextToken :: Lens' ListFindingsResponse (Maybe Text)

A token used for pagination of results returned.

listFindingsResponse_httpStatus :: Lens' ListFindingsResponse Int

The response's http status code.

listFindingsResponse_findings :: Lens' ListFindingsResponse [FindingSummary]

A list of findings retrieved from the analyzer that match the filter criteria specified, if any.