Amazonka.AccessAnalyzer.StartPolicyGeneration

Description

Starts the policy generation request.

Creating a Request

data StartPolicyGeneration

See: newStartPolicyGeneration smart constructor.

Constructors

StartPolicyGeneration'

Fields

clientToken :: Maybe Text

A unique, case-sensitive identifier that you provide to ensure the idempotency of the request. Idempotency ensures that an API request completes only once. With an idempotent request, if the original request completes successfully, the subsequent retries with the same client token return the result from the original successful request and they have no additional effect.

If you do not specify a client token, one is automatically generated by the Amazon Web Services SDK.

cloudTrailDetails :: Maybe CloudTrailDetails

A CloudTrailDetails object that contains details about a Trail that you want to analyze to generate policies.

policyGenerationDetails :: PolicyGenerationDetails

Contains the ARN of the IAM entity (user or role) for which you are generating a policy.

newStartPolicyGeneration

Arguments

:: PolicyGenerationDetails
$sel:policyGenerationDetails:StartPolicyGeneration'

-> StartPolicyGeneration

Create a value of StartPolicyGeneration with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:clientToken:StartPolicyGeneration', startPolicyGeneration_clientToken - A unique, case-sensitive identifier that you provide to ensure the idempotency of the request. Idempotency ensures that an API request completes only once. With an idempotent request, if the original request completes successfully, the subsequent retries with the same client token return the result from the original successful request and they have no additional effect.

If you do not specify a client token, one is automatically generated by the Amazon Web Services SDK.

$sel:cloudTrailDetails:StartPolicyGeneration', startPolicyGeneration_cloudTrailDetails - A CloudTrailDetails object that contains details about a Trail that you want to analyze to generate policies.

$sel:policyGenerationDetails:StartPolicyGeneration', startPolicyGeneration_policyGenerationDetails - Contains the ARN of the IAM entity (user or role) for which you are generating a policy.

Request Lenses

startPolicyGeneration_clientToken :: Lens' StartPolicyGeneration (Maybe Text)

A unique, case-sensitive identifier that you provide to ensure the idempotency of the request. Idempotency ensures that an API request completes only once. With an idempotent request, if the original request completes successfully, the subsequent retries with the same client token return the result from the original successful request and they have no additional effect.

If you do not specify a client token, one is automatically generated by the Amazon Web Services SDK.

startPolicyGeneration_cloudTrailDetails :: Lens' StartPolicyGeneration (Maybe CloudTrailDetails)

A CloudTrailDetails object that contains details about a Trail that you want to analyze to generate policies.

startPolicyGeneration_policyGenerationDetails :: Lens' StartPolicyGeneration PolicyGenerationDetails

Contains the ARN of the IAM entity (user or role) for which you are generating a policy.

Destructuring the Response

data StartPolicyGenerationResponse

See: newStartPolicyGenerationResponse smart constructor.

Constructors

StartPolicyGenerationResponse'

Fields

httpStatus :: Int

The response's http status code.

jobId :: Text

The JobId that is returned by the StartPolicyGeneration operation. The JobId can be used with GetGeneratedPolicy to retrieve the generated policies or used with CancelPolicyGeneration to cancel the policy generation request.

newStartPolicyGenerationResponse

Arguments

:: Int
$sel:httpStatus:StartPolicyGenerationResponse'

-> Text
$sel:jobId:StartPolicyGenerationResponse'

-> StartPolicyGenerationResponse

Create a value of StartPolicyGenerationResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:httpStatus:StartPolicyGenerationResponse', startPolicyGenerationResponse_httpStatus - The response's http status code.

$sel:jobId:StartPolicyGenerationResponse', startPolicyGenerationResponse_jobId - The JobId that is returned by the StartPolicyGeneration operation. The JobId can be used with GetGeneratedPolicy to retrieve the generated policies or used with CancelPolicyGeneration to cancel the policy generation request.

Response Lenses

startPolicyGenerationResponse_httpStatus :: Lens' StartPolicyGenerationResponse Int

The response's http status code.

startPolicyGenerationResponse_jobId :: Lens' StartPolicyGenerationResponse Text

The JobId that is returned by the StartPolicyGeneration operation. The JobId can be used with GetGeneratedPolicy to retrieve the generated policies or used with CancelPolicyGeneration to cancel the policy generation request.