Amazonka.AccessAnalyzer.StartResourceScan

Description

Immediately starts a scan of the policies applied to the specified resource.

Creating a Request

data StartResourceScan

Starts a scan of the policies applied to the specified resource.

See: newStartResourceScan smart constructor.

Constructors

StartResourceScan'

Fields

analyzerArn :: Text

The ARN of the analyzer to use to scan the policies applied to the specified resource.

resourceArn :: Text

The ARN of the resource to scan.

newStartResourceScan

Arguments

:: Text
$sel:analyzerArn:StartResourceScan'

-> Text
$sel:resourceArn:StartResourceScan'

-> StartResourceScan

Create a value of StartResourceScan with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:analyzerArn:StartResourceScan', startResourceScan_analyzerArn - The ARN of the analyzer to use to scan the policies applied to the specified resource.

$sel:resourceArn:StartResourceScan', startResourceScan_resourceArn - The ARN of the resource to scan.

Request Lenses

startResourceScan_analyzerArn :: Lens' StartResourceScan Text

The ARN of the analyzer to use to scan the policies applied to the specified resource.

startResourceScan_resourceArn :: Lens' StartResourceScan Text

The ARN of the resource to scan.

Destructuring the Response

data StartResourceScanResponse

See: newStartResourceScanResponse smart constructor.

Constructors

StartResourceScanResponse'

newStartResourceScanResponse :: StartResourceScanResponse

Create a value of StartResourceScanResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.