Amazonka.AccessAnalyzer.Types.AccessPreviewFinding

Description

Documentation

data AccessPreviewFinding

An access preview finding generated by the access preview.

See: newAccessPreviewFinding smart constructor.

Constructors

AccessPreviewFinding'

Fields

existingFindingStatus :: Maybe FindingStatus

The existing status of the finding, provided only for existing findings.

error :: Maybe Text

An error.

isPublic :: Maybe Bool

Indicates whether the policy that generated the finding allows public access to the resource.

action :: Maybe [Text]

The action in the analyzed policy statement that an external principal has permission to perform.

sources :: Maybe [FindingSource]

The sources of the finding. This indicates how the access that generated the finding is granted. It is populated for Amazon S3 bucket findings.

resource :: Maybe Text

The resource that an external principal has access to. This is the resource associated with the access preview.

principal :: Maybe (HashMap Text Text)

The external principal that has access to a resource within the zone of trust.

existingFindingId :: Maybe Text

The existing ID of the finding in IAM Access Analyzer, provided only for existing findings.

condition :: Maybe (HashMap Text Text)

The condition in the analyzed policy statement that resulted in a finding.

changeType :: FindingChangeType

Provides context on how the access preview finding compares to existing access identified in IAM Access Analyzer.

New - The finding is for newly-introduced access.

Unchanged - The preview finding is an existing finding that would remain unchanged.

Changed - The preview finding is an existing finding with a change in status.

For example, a Changed finding with preview status Resolved and existing status Active indicates the existing Active finding would become Resolved as a result of the proposed permissions change.

createdAt :: POSIX

The time at which the access preview finding was created.

id :: Text

The ID of the access preview finding. This ID uniquely identifies the element in the list of access preview findings and is not related to the finding ID in Access Analyzer.

resourceOwnerAccount :: Text

The Amazon Web Services account ID that owns the resource. For most Amazon Web Services resources, the owning account is the account in which the resource was created.

resourceType :: ResourceType

The type of the resource that can be accessed in the finding.

status :: FindingStatus

The preview status of the finding. This is what the status of the finding would be after permissions deployment. For example, a Changed finding with preview status Resolved and existing status Active indicates the existing Active finding would become Resolved as a result of the proposed permissions change.

newAccessPreviewFinding

Arguments

:: FindingChangeType
$sel:changeType:AccessPreviewFinding'

-> UTCTime
$sel:createdAt:AccessPreviewFinding'

-> Text
$sel:id:AccessPreviewFinding'

-> Text
$sel:resourceOwnerAccount:AccessPreviewFinding'

-> ResourceType
$sel:resourceType:AccessPreviewFinding'

-> FindingStatus
$sel:status:AccessPreviewFinding'

-> AccessPreviewFinding

Create a value of AccessPreviewFinding with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:existingFindingStatus:AccessPreviewFinding', accessPreviewFinding_existingFindingStatus - The existing status of the finding, provided only for existing findings.

$sel:error:AccessPreviewFinding', accessPreviewFinding_error - An error.

$sel:isPublic:AccessPreviewFinding', accessPreviewFinding_isPublic - Indicates whether the policy that generated the finding allows public access to the resource.

$sel:action:AccessPreviewFinding', accessPreviewFinding_action - The action in the analyzed policy statement that an external principal has permission to perform.

$sel:sources:AccessPreviewFinding', accessPreviewFinding_sources - The sources of the finding. This indicates how the access that generated the finding is granted. It is populated for Amazon S3 bucket findings.

$sel:resource:AccessPreviewFinding', accessPreviewFinding_resource - The resource that an external principal has access to. This is the resource associated with the access preview.

$sel:principal:AccessPreviewFinding', accessPreviewFinding_principal - The external principal that has access to a resource within the zone of trust.

$sel:existingFindingId:AccessPreviewFinding', accessPreviewFinding_existingFindingId - The existing ID of the finding in IAM Access Analyzer, provided only for existing findings.

$sel:condition:AccessPreviewFinding', accessPreviewFinding_condition - The condition in the analyzed policy statement that resulted in a finding.

$sel:changeType:AccessPreviewFinding', accessPreviewFinding_changeType - Provides context on how the access preview finding compares to existing access identified in IAM Access Analyzer.

New - The finding is for newly-introduced access.

Unchanged - The preview finding is an existing finding that would remain unchanged.

Changed - The preview finding is an existing finding with a change in status.

For example, a Changed finding with preview status Resolved and existing status Active indicates the existing Active finding would become Resolved as a result of the proposed permissions change.

$sel:createdAt:AccessPreviewFinding', accessPreviewFinding_createdAt - The time at which the access preview finding was created.

$sel:id:AccessPreviewFinding', accessPreviewFinding_id - The ID of the access preview finding. This ID uniquely identifies the element in the list of access preview findings and is not related to the finding ID in Access Analyzer.

$sel:resourceOwnerAccount:AccessPreviewFinding', accessPreviewFinding_resourceOwnerAccount - The Amazon Web Services account ID that owns the resource. For most Amazon Web Services resources, the owning account is the account in which the resource was created.

$sel:resourceType:AccessPreviewFinding', accessPreviewFinding_resourceType - The type of the resource that can be accessed in the finding.

$sel:status:AccessPreviewFinding', accessPreviewFinding_status - The preview status of the finding. This is what the status of the finding would be after permissions deployment. For example, a Changed finding with preview status Resolved and existing status Active indicates the existing Active finding would become Resolved as a result of the proposed permissions change.

accessPreviewFinding_existingFindingStatus :: Lens' AccessPreviewFinding (Maybe FindingStatus)

The existing status of the finding, provided only for existing findings.

accessPreviewFinding_error :: Lens' AccessPreviewFinding (Maybe Text)

An error.

accessPreviewFinding_isPublic :: Lens' AccessPreviewFinding (Maybe Bool)

Indicates whether the policy that generated the finding allows public access to the resource.

accessPreviewFinding_action :: Lens' AccessPreviewFinding (Maybe [Text])

The action in the analyzed policy statement that an external principal has permission to perform.

accessPreviewFinding_sources :: Lens' AccessPreviewFinding (Maybe [FindingSource])

The sources of the finding. This indicates how the access that generated the finding is granted. It is populated for Amazon S3 bucket findings.

accessPreviewFinding_resource :: Lens' AccessPreviewFinding (Maybe Text)

The resource that an external principal has access to. This is the resource associated with the access preview.

accessPreviewFinding_principal :: Lens' AccessPreviewFinding (Maybe (HashMap Text Text))

The external principal that has access to a resource within the zone of trust.

accessPreviewFinding_existingFindingId :: Lens' AccessPreviewFinding (Maybe Text)

The existing ID of the finding in IAM Access Analyzer, provided only for existing findings.

accessPreviewFinding_condition :: Lens' AccessPreviewFinding (Maybe (HashMap Text Text))

The condition in the analyzed policy statement that resulted in a finding.

accessPreviewFinding_changeType :: Lens' AccessPreviewFinding FindingChangeType

Provides context on how the access preview finding compares to existing access identified in IAM Access Analyzer.

New - The finding is for newly-introduced access.

Unchanged - The preview finding is an existing finding that would remain unchanged.

Changed - The preview finding is an existing finding with a change in status.

For example, a Changed finding with preview status Resolved and existing status Active indicates the existing Active finding would become Resolved as a result of the proposed permissions change.

accessPreviewFinding_createdAt :: Lens' AccessPreviewFinding UTCTime

The time at which the access preview finding was created.

accessPreviewFinding_id :: Lens' AccessPreviewFinding Text

The ID of the access preview finding. This ID uniquely identifies the element in the list of access preview findings and is not related to the finding ID in Access Analyzer.

accessPreviewFinding_resourceOwnerAccount :: Lens' AccessPreviewFinding Text

The Amazon Web Services account ID that owns the resource. For most Amazon Web Services resources, the owning account is the account in which the resource was created.

accessPreviewFinding_resourceType :: Lens' AccessPreviewFinding ResourceType

The type of the resource that can be accessed in the finding.

accessPreviewFinding_status :: Lens' AccessPreviewFinding FindingStatus

The preview status of the finding. This is what the status of the finding would be after permissions deployment. For example, a Changed finding with preview status Resolved and existing status Active indicates the existing Active finding would become Resolved as a result of the proposed permissions change.