Amazonka.AccessAnalyzer.Types.AnalyzedResource

Description

Documentation

data AnalyzedResource

Contains details about the analyzed resource.

See: newAnalyzedResource smart constructor.

Constructors

AnalyzedResource'

Fields

status :: Maybe FindingStatus

The current status of the finding generated from the analyzed resource.

actions :: Maybe [Text]

The actions that an external principal is granted permission to use by the policy that generated the finding.

error :: Maybe Text

An error message.

sharedVia :: Maybe [Text]

Indicates how the access that generated the finding is granted. This is populated for Amazon S3 bucket findings.

analyzedAt :: POSIX

The time at which the resource was analyzed.

createdAt :: POSIX

The time at which the finding was created.

isPublic :: Bool

Indicates whether the policy that generated the finding grants public access to the resource.

resourceArn :: Text

The ARN of the resource that was analyzed.

resourceOwnerAccount :: Text

The Amazon Web Services account ID that owns the resource.

resourceType :: ResourceType

The type of the resource that was analyzed.

updatedAt :: POSIX

The time at which the finding was updated.

newAnalyzedResource

Arguments

:: UTCTime
$sel:analyzedAt:AnalyzedResource'

-> UTCTime
$sel:createdAt:AnalyzedResource'

-> Bool
$sel:isPublic:AnalyzedResource'

-> Text
$sel:resourceArn:AnalyzedResource'

-> Text
$sel:resourceOwnerAccount:AnalyzedResource'

-> ResourceType
$sel:resourceType:AnalyzedResource'

-> UTCTime
$sel:updatedAt:AnalyzedResource'

-> AnalyzedResource

Create a value of AnalyzedResource with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:status:AnalyzedResource', analyzedResource_status - The current status of the finding generated from the analyzed resource.

$sel:actions:AnalyzedResource', analyzedResource_actions - The actions that an external principal is granted permission to use by the policy that generated the finding.

$sel:error:AnalyzedResource', analyzedResource_error - An error message.

$sel:sharedVia:AnalyzedResource', analyzedResource_sharedVia - Indicates how the access that generated the finding is granted. This is populated for Amazon S3 bucket findings.

$sel:analyzedAt:AnalyzedResource', analyzedResource_analyzedAt - The time at which the resource was analyzed.

$sel:createdAt:AnalyzedResource', analyzedResource_createdAt - The time at which the finding was created.

$sel:isPublic:AnalyzedResource', analyzedResource_isPublic - Indicates whether the policy that generated the finding grants public access to the resource.

$sel:resourceArn:AnalyzedResource', analyzedResource_resourceArn - The ARN of the resource that was analyzed.

$sel:resourceOwnerAccount:AnalyzedResource', analyzedResource_resourceOwnerAccount - The Amazon Web Services account ID that owns the resource.

$sel:resourceType:AnalyzedResource', analyzedResource_resourceType - The type of the resource that was analyzed.

$sel:updatedAt:AnalyzedResource', analyzedResource_updatedAt - The time at which the finding was updated.

analyzedResource_status :: Lens' AnalyzedResource (Maybe FindingStatus)

The current status of the finding generated from the analyzed resource.

analyzedResource_actions :: Lens' AnalyzedResource (Maybe [Text])

The actions that an external principal is granted permission to use by the policy that generated the finding.

analyzedResource_error :: Lens' AnalyzedResource (Maybe Text)

An error message.

analyzedResource_sharedVia :: Lens' AnalyzedResource (Maybe [Text])

Indicates how the access that generated the finding is granted. This is populated for Amazon S3 bucket findings.

analyzedResource_analyzedAt :: Lens' AnalyzedResource UTCTime

The time at which the resource was analyzed.

analyzedResource_createdAt :: Lens' AnalyzedResource UTCTime

The time at which the finding was created.

analyzedResource_isPublic :: Lens' AnalyzedResource Bool

Indicates whether the policy that generated the finding grants public access to the resource.

analyzedResource_resourceArn :: Lens' AnalyzedResource Text

The ARN of the resource that was analyzed.

analyzedResource_resourceOwnerAccount :: Lens' AnalyzedResource Text

The Amazon Web Services account ID that owns the resource.

analyzedResource_resourceType :: Lens' AnalyzedResource ResourceType

The type of the resource that was analyzed.

analyzedResource_updatedAt :: Lens' AnalyzedResource UTCTime

The time at which the finding was updated.