Amazonka.AccessAnalyzer.Types.Finding

Description

Documentation

data Finding

Contains information about a finding.

See: newFinding smart constructor.

Constructors

Finding'

Fields

error :: Maybe Text

An error.

isPublic :: Maybe Bool

Indicates whether the policy that generated the finding allows public access to the resource.

action :: Maybe [Text]

The action in the analyzed policy statement that an external principal has permission to use.

sources :: Maybe [FindingSource]

The sources of the finding. This indicates how the access that generated the finding is granted. It is populated for Amazon S3 bucket findings.

resource :: Maybe Text

The resource that an external principal has access to.

principal :: Maybe (HashMap Text Text)

The external principal that access to a resource within the zone of trust.

analyzedAt :: POSIX

The time at which the resource was analyzed.

condition :: HashMap Text Text

The condition in the analyzed policy statement that resulted in a finding.

createdAt :: POSIX

The time at which the finding was generated.

id :: Text

The ID of the finding.

resourceOwnerAccount :: Text

The Amazon Web Services account ID that owns the resource.

resourceType :: ResourceType

The type of the resource identified in the finding.

status :: FindingStatus

The current status of the finding.

updatedAt :: POSIX

The time at which the finding was updated.

newFinding

Arguments

:: UTCTime
$sel:analyzedAt:Finding'

-> UTCTime
$sel:createdAt:Finding'

-> Text
$sel:id:Finding'

-> Text
$sel:resourceOwnerAccount:Finding'

-> ResourceType
$sel:resourceType:Finding'

-> FindingStatus
$sel:status:Finding'

-> UTCTime
$sel:updatedAt:Finding'

-> Finding

Create a value of Finding with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:error:Finding', finding_error - An error.

$sel:isPublic:Finding', finding_isPublic - Indicates whether the policy that generated the finding allows public access to the resource.

$sel:action:Finding', finding_action - The action in the analyzed policy statement that an external principal has permission to use.

$sel:sources:Finding', finding_sources - The sources of the finding. This indicates how the access that generated the finding is granted. It is populated for Amazon S3 bucket findings.

$sel:resource:Finding', finding_resource - The resource that an external principal has access to.

$sel:principal:Finding', finding_principal - The external principal that access to a resource within the zone of trust.

$sel:analyzedAt:Finding', finding_analyzedAt - The time at which the resource was analyzed.

$sel:condition:Finding', finding_condition - The condition in the analyzed policy statement that resulted in a finding.

$sel:createdAt:Finding', finding_createdAt - The time at which the finding was generated.

$sel:id:Finding', finding_id - The ID of the finding.

$sel:resourceOwnerAccount:Finding', finding_resourceOwnerAccount - The Amazon Web Services account ID that owns the resource.

$sel:resourceType:Finding', finding_resourceType - The type of the resource identified in the finding.

$sel:status:Finding', finding_status - The current status of the finding.

$sel:updatedAt:Finding', finding_updatedAt - The time at which the finding was updated.

finding_error :: Lens' Finding (Maybe Text)

An error.

finding_isPublic :: Lens' Finding (Maybe Bool)

Indicates whether the policy that generated the finding allows public access to the resource.

finding_action :: Lens' Finding (Maybe [Text])

The action in the analyzed policy statement that an external principal has permission to use.

finding_sources :: Lens' Finding (Maybe [FindingSource])

The sources of the finding. This indicates how the access that generated the finding is granted. It is populated for Amazon S3 bucket findings.

finding_resource :: Lens' Finding (Maybe Text)

The resource that an external principal has access to.

finding_principal :: Lens' Finding (Maybe (HashMap Text Text))

The external principal that access to a resource within the zone of trust.

finding_analyzedAt :: Lens' Finding UTCTime

The time at which the resource was analyzed.

finding_condition :: Lens' Finding (HashMap Text Text)

The condition in the analyzed policy statement that resulted in a finding.

finding_createdAt :: Lens' Finding UTCTime

The time at which the finding was generated.

finding_id :: Lens' Finding Text

The ID of the finding.

finding_resourceOwnerAccount :: Lens' Finding Text

The Amazon Web Services account ID that owns the resource.

finding_resourceType :: Lens' Finding ResourceType

The type of the resource identified in the finding.

finding_status :: Lens' Finding FindingStatus

The current status of the finding.

finding_updatedAt :: Lens' Finding UTCTime

The time at which the finding was updated.