Amazonka.AccessAnalyzer.Types.KmsGrantConfiguration

Description

Documentation

data KmsGrantConfiguration

A proposed grant configuration for a KMS key. For more information, see CreateGrant.

See: newKmsGrantConfiguration smart constructor.

Constructors

KmsGrantConfiguration'

Fields

retiringPrincipal :: Maybe Text

The principal that is given permission to retire the grant by using RetireGrant operation.

constraints :: Maybe KmsGrantConstraints

Use this structure to propose allowing cryptographic operations in the grant only when the operation request includes the specified encryption context.

granteePrincipal :: Text

The principal that is given permission to perform the operations that the grant permits.

issuingAccount :: Text

The Amazon Web Services account under which the grant was issued. The account is used to propose KMS grants issued by accounts other than the owner of the key.

operations :: [KmsGrantOperation]

A list of operations that the grant permits.

newKmsGrantConfiguration

Arguments

:: Text
$sel:granteePrincipal:KmsGrantConfiguration'

-> Text
$sel:issuingAccount:KmsGrantConfiguration'

-> KmsGrantConfiguration

Create a value of KmsGrantConfiguration with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:retiringPrincipal:KmsGrantConfiguration', kmsGrantConfiguration_retiringPrincipal - The principal that is given permission to retire the grant by using RetireGrant operation.

$sel:constraints:KmsGrantConfiguration', kmsGrantConfiguration_constraints - Use this structure to propose allowing cryptographic operations in the grant only when the operation request includes the specified encryption context.

$sel:granteePrincipal:KmsGrantConfiguration', kmsGrantConfiguration_granteePrincipal - The principal that is given permission to perform the operations that the grant permits.

$sel:issuingAccount:KmsGrantConfiguration', kmsGrantConfiguration_issuingAccount - The Amazon Web Services account under which the grant was issued. The account is used to propose KMS grants issued by accounts other than the owner of the key.

$sel:operations:KmsGrantConfiguration', kmsGrantConfiguration_operations - A list of operations that the grant permits.

kmsGrantConfiguration_retiringPrincipal :: Lens' KmsGrantConfiguration (Maybe Text)

The principal that is given permission to retire the grant by using RetireGrant operation.

kmsGrantConfiguration_constraints :: Lens' KmsGrantConfiguration (Maybe KmsGrantConstraints)

Use this structure to propose allowing cryptographic operations in the grant only when the operation request includes the specified encryption context.

kmsGrantConfiguration_granteePrincipal :: Lens' KmsGrantConfiguration Text

The principal that is given permission to perform the operations that the grant permits.

kmsGrantConfiguration_issuingAccount :: Lens' KmsGrantConfiguration Text

The Amazon Web Services account under which the grant was issued. The account is used to propose KMS grants issued by accounts other than the owner of the key.

kmsGrantConfiguration_operations :: Lens' KmsGrantConfiguration [KmsGrantOperation]

A list of operations that the grant permits.