Amazonka.AccessAnalyzer.Types.NetworkOriginConfiguration

Description

Documentation

data NetworkOriginConfiguration

The proposed InternetConfiguration or VpcConfiguration to apply to the Amazon S3 access point. VpcConfiguration does not apply to multi-region access points. You can make the access point accessible from the internet, or you can specify that all requests made through that access point must originate from a specific virtual private cloud (VPC). You can specify only one type of network configuration. For more information, see Creating access points.

See: newNetworkOriginConfiguration smart constructor.

Constructors

NetworkOriginConfiguration'

Fields

internetConfiguration :: Maybe InternetConfiguration

The configuration for the Amazon S3 access point or multi-region access point with an Internet origin.

vpcConfiguration :: Maybe VpcConfiguration

newNetworkOriginConfiguration :: NetworkOriginConfiguration

Create a value of NetworkOriginConfiguration with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:internetConfiguration:NetworkOriginConfiguration', networkOriginConfiguration_internetConfiguration - The configuration for the Amazon S3 access point or multi-region access point with an Internet origin.

$sel:vpcConfiguration:NetworkOriginConfiguration', networkOriginConfiguration_vpcConfiguration - Undocumented member.

networkOriginConfiguration_internetConfiguration :: Lens' NetworkOriginConfiguration (Maybe InternetConfiguration)

The configuration for the Amazon S3 access point or multi-region access point with an Internet origin.

networkOriginConfiguration_vpcConfiguration :: Lens' NetworkOriginConfiguration (Maybe VpcConfiguration)

Undocumented member.