Amazonka.AccessAnalyzer.Types.PolicyGeneration

Description

Documentation

data PolicyGeneration

Contains details about the policy generation status and properties.

See: newPolicyGeneration smart constructor.

Constructors

PolicyGeneration'

Fields

completedOn :: Maybe POSIX

A timestamp of when the policy generation was completed.

jobId :: Text

The JobId that is returned by the StartPolicyGeneration operation. The JobId can be used with GetGeneratedPolicy to retrieve the generated policies or used with CancelPolicyGeneration to cancel the policy generation request.

principalArn :: Text

The ARN of the IAM entity (user or role) for which you are generating a policy.

startedOn :: POSIX

A timestamp of when the policy generation started.

status :: JobStatus

The status of the policy generation request.

newPolicyGeneration

Arguments

:: Text
$sel:jobId:PolicyGeneration'

-> Text
$sel:principalArn:PolicyGeneration'

-> UTCTime
$sel:startedOn:PolicyGeneration'

-> JobStatus
$sel:status:PolicyGeneration'

-> PolicyGeneration

Create a value of PolicyGeneration with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:completedOn:PolicyGeneration', policyGeneration_completedOn - A timestamp of when the policy generation was completed.

$sel:jobId:PolicyGeneration', policyGeneration_jobId - The JobId that is returned by the StartPolicyGeneration operation. The JobId can be used with GetGeneratedPolicy to retrieve the generated policies or used with CancelPolicyGeneration to cancel the policy generation request.

$sel:principalArn:PolicyGeneration', policyGeneration_principalArn - The ARN of the IAM entity (user or role) for which you are generating a policy.

$sel:startedOn:PolicyGeneration', policyGeneration_startedOn - A timestamp of when the policy generation started.

$sel:status:PolicyGeneration', policyGeneration_status - The status of the policy generation request.

policyGeneration_completedOn :: Lens' PolicyGeneration (Maybe UTCTime)

A timestamp of when the policy generation was completed.

policyGeneration_jobId :: Lens' PolicyGeneration Text

The JobId that is returned by the StartPolicyGeneration operation. The JobId can be used with GetGeneratedPolicy to retrieve the generated policies or used with CancelPolicyGeneration to cancel the policy generation request.

policyGeneration_principalArn :: Lens' PolicyGeneration Text

The ARN of the IAM entity (user or role) for which you are generating a policy.

policyGeneration_startedOn :: Lens' PolicyGeneration UTCTime

A timestamp of when the policy generation started.

policyGeneration_status :: Lens' PolicyGeneration JobStatus

The status of the policy generation request.