Amazonka.AccessAnalyzer.Types.S3BucketAclGrantConfiguration

Description

Documentation

data S3BucketAclGrantConfiguration

A proposed access control list grant configuration for an Amazon S3 bucket. For more information, see How to Specify an ACL.

See: newS3BucketAclGrantConfiguration smart constructor.

Constructors

S3BucketAclGrantConfiguration'

Fields

grantee :: AclGrantee

The grantee to whom you’re assigning access rights.

permission :: AclPermission

The permissions being granted.

newS3BucketAclGrantConfiguration

Arguments

:: AclGrantee
$sel:grantee:S3BucketAclGrantConfiguration'

-> AclPermission
$sel:permission:S3BucketAclGrantConfiguration'

-> S3BucketAclGrantConfiguration

Create a value of S3BucketAclGrantConfiguration with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:grantee:S3BucketAclGrantConfiguration', s3BucketAclGrantConfiguration_grantee - The grantee to whom you’re assigning access rights.

$sel:permission:S3BucketAclGrantConfiguration', s3BucketAclGrantConfiguration_permission - The permissions being granted.

s3BucketAclGrantConfiguration_grantee :: Lens' S3BucketAclGrantConfiguration AclGrantee

The grantee to whom you’re assigning access rights.

s3BucketAclGrantConfiguration_permission :: Lens' S3BucketAclGrantConfiguration AclPermission

The permissions being granted.