Amazonka.AccessAnalyzer.Types.S3BucketConfiguration

Description

Documentation

data S3BucketConfiguration

Proposed access control configuration for an Amazon S3 bucket. You can propose a configuration for a new Amazon S3 bucket or an existing Amazon S3 bucket that you own by specifying the Amazon S3 bucket policy, bucket ACLs, bucket BPA settings, Amazon S3 access points, and multi-region access points attached to the bucket. If the configuration is for an existing Amazon S3 bucket and you do not specify the Amazon S3 bucket policy, the access preview uses the existing policy attached to the bucket. If the access preview is for a new resource and you do not specify the Amazon S3 bucket policy, the access preview assumes a bucket without a policy. To propose deletion of an existing bucket policy, you can specify an empty string. For more information about bucket policy limits, see Bucket Policy Examples.

See: newS3BucketConfiguration smart constructor.

Constructors

S3BucketConfiguration'

Fields

accessPoints :: Maybe (HashMap Text S3AccessPointConfiguration)

The configuration of Amazon S3 access points or multi-region access points for the bucket. You can propose up to 10 new access points per bucket.

bucketPublicAccessBlock :: Maybe S3PublicAccessBlockConfiguration

The proposed block public access configuration for the Amazon S3 bucket.

bucketAclGrants :: Maybe [S3BucketAclGrantConfiguration]

The proposed list of ACL grants for the Amazon S3 bucket. You can propose up to 100 ACL grants per bucket. If the proposed grant configuration is for an existing bucket, the access preview uses the proposed list of grant configurations in place of the existing grants. Otherwise, the access preview uses the existing grants for the bucket.

bucketPolicy :: Maybe Text

The proposed bucket policy for the Amazon S3 bucket.

newS3BucketConfiguration :: S3BucketConfiguration

Create a value of S3BucketConfiguration with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:accessPoints:S3BucketConfiguration', s3BucketConfiguration_accessPoints - The configuration of Amazon S3 access points or multi-region access points for the bucket. You can propose up to 10 new access points per bucket.

$sel:bucketPublicAccessBlock:S3BucketConfiguration', s3BucketConfiguration_bucketPublicAccessBlock - The proposed block public access configuration for the Amazon S3 bucket.

$sel:bucketAclGrants:S3BucketConfiguration', s3BucketConfiguration_bucketAclGrants - The proposed list of ACL grants for the Amazon S3 bucket. You can propose up to 100 ACL grants per bucket. If the proposed grant configuration is for an existing bucket, the access preview uses the proposed list of grant configurations in place of the existing grants. Otherwise, the access preview uses the existing grants for the bucket.

$sel:bucketPolicy:S3BucketConfiguration', s3BucketConfiguration_bucketPolicy - The proposed bucket policy for the Amazon S3 bucket.

s3BucketConfiguration_accessPoints :: Lens' S3BucketConfiguration (Maybe (HashMap Text S3AccessPointConfiguration))

The configuration of Amazon S3 access points or multi-region access points for the bucket. You can propose up to 10 new access points per bucket.

s3BucketConfiguration_bucketPublicAccessBlock :: Lens' S3BucketConfiguration (Maybe S3PublicAccessBlockConfiguration)

The proposed block public access configuration for the Amazon S3 bucket.

s3BucketConfiguration_bucketAclGrants :: Lens' S3BucketConfiguration (Maybe [S3BucketAclGrantConfiguration])

The proposed list of ACL grants for the Amazon S3 bucket. You can propose up to 100 ACL grants per bucket. If the proposed grant configuration is for an existing bucket, the access preview uses the proposed list of grant configurations in place of the existing grants. Otherwise, the access preview uses the existing grants for the bucket.

s3BucketConfiguration_bucketPolicy :: Lens' S3BucketConfiguration (Maybe Text)

The proposed bucket policy for the Amazon S3 bucket.