Amazonka.AccessAnalyzer.Types.S3PublicAccessBlockConfiguration

Description

Documentation

data S3PublicAccessBlockConfiguration

The PublicAccessBlock configuration to apply to this Amazon S3 bucket. If the proposed configuration is for an existing Amazon S3 bucket and the configuration is not specified, the access preview uses the existing setting. If the proposed configuration is for a new bucket and the configuration is not specified, the access preview uses false. If the proposed configuration is for a new access point or multi-region access point and the access point BPA configuration is not specified, the access preview uses true. For more information, see PublicAccessBlockConfiguration.

See: newS3PublicAccessBlockConfiguration smart constructor.

Constructors

S3PublicAccessBlockConfiguration'

Fields

ignorePublicAcls :: Bool

Specifies whether Amazon S3 should ignore public ACLs for this bucket and objects in this bucket.

restrictPublicBuckets :: Bool

Specifies whether Amazon S3 should restrict public bucket policies for this bucket.

newS3PublicAccessBlockConfiguration

Arguments

:: Bool
$sel:ignorePublicAcls:S3PublicAccessBlockConfiguration'

-> Bool
$sel:restrictPublicBuckets:S3PublicAccessBlockConfiguration'

-> S3PublicAccessBlockConfiguration

Create a value of S3PublicAccessBlockConfiguration with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:ignorePublicAcls:S3PublicAccessBlockConfiguration', s3PublicAccessBlockConfiguration_ignorePublicAcls - Specifies whether Amazon S3 should ignore public ACLs for this bucket and objects in this bucket.

$sel:restrictPublicBuckets:S3PublicAccessBlockConfiguration', s3PublicAccessBlockConfiguration_restrictPublicBuckets - Specifies whether Amazon S3 should restrict public bucket policies for this bucket.

s3PublicAccessBlockConfiguration_ignorePublicAcls :: Lens' S3PublicAccessBlockConfiguration Bool

Specifies whether Amazon S3 should ignore public ACLs for this bucket and objects in this bucket.

s3PublicAccessBlockConfiguration_restrictPublicBuckets :: Lens' S3PublicAccessBlockConfiguration Bool

Specifies whether Amazon S3 should restrict public bucket policies for this bucket.