Amazonka.AccessAnalyzer.Types.ValidatePolicyFinding

Description

Documentation

data ValidatePolicyFinding

A finding in a policy. Each finding is an actionable recommendation that can be used to improve the policy.

See: newValidatePolicyFinding smart constructor.

Constructors

ValidatePolicyFinding'

Fields

findingDetails :: Text

A localized message that explains the finding and provides guidance on how to address it.

findingType :: ValidatePolicyFindingType

The impact of the finding.

Security warnings report when the policy allows access that we consider overly permissive.

Errors report when a part of the policy is not functional.

Warnings report non-security issues when a policy does not conform to policy writing best practices.

Suggestions recommend stylistic improvements in the policy that do not impact access.

issueCode :: Text

The issue code provides an identifier of the issue associated with this finding.

learnMoreLink :: Text

A link to additional documentation about the type of finding.

locations :: [Location]

The list of locations in the policy document that are related to the finding. The issue code provides a summary of an issue identified by the finding.

newValidatePolicyFinding

Arguments

:: Text
$sel:findingDetails:ValidatePolicyFinding'

-> ValidatePolicyFindingType
$sel:findingType:ValidatePolicyFinding'

-> Text
$sel:issueCode:ValidatePolicyFinding'

-> Text
$sel:learnMoreLink:ValidatePolicyFinding'

-> ValidatePolicyFinding

Create a value of ValidatePolicyFinding with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:findingDetails:ValidatePolicyFinding', validatePolicyFinding_findingDetails - A localized message that explains the finding and provides guidance on how to address it.

$sel:findingType:ValidatePolicyFinding', validatePolicyFinding_findingType - The impact of the finding.

Security warnings report when the policy allows access that we consider overly permissive.

Errors report when a part of the policy is not functional.

Warnings report non-security issues when a policy does not conform to policy writing best practices.

Suggestions recommend stylistic improvements in the policy that do not impact access.

$sel:issueCode:ValidatePolicyFinding', validatePolicyFinding_issueCode - The issue code provides an identifier of the issue associated with this finding.

$sel:learnMoreLink:ValidatePolicyFinding', validatePolicyFinding_learnMoreLink - A link to additional documentation about the type of finding.

$sel:locations:ValidatePolicyFinding', validatePolicyFinding_locations - The list of locations in the policy document that are related to the finding. The issue code provides a summary of an issue identified by the finding.

validatePolicyFinding_findingDetails :: Lens' ValidatePolicyFinding Text

A localized message that explains the finding and provides guidance on how to address it.

validatePolicyFinding_findingType :: Lens' ValidatePolicyFinding ValidatePolicyFindingType

The impact of the finding.

Security warnings report when the policy allows access that we consider overly permissive.

Errors report when a part of the policy is not functional.

Warnings report non-security issues when a policy does not conform to policy writing best practices.

Suggestions recommend stylistic improvements in the policy that do not impact access.

validatePolicyFinding_issueCode :: Lens' ValidatePolicyFinding Text

The issue code provides an identifier of the issue associated with this finding.

validatePolicyFinding_learnMoreLink :: Lens' ValidatePolicyFinding Text

A link to additional documentation about the type of finding.

validatePolicyFinding_locations :: Lens' ValidatePolicyFinding [Location]

The list of locations in the policy document that are related to the finding. The issue code provides a summary of an issue identified by the finding.