Amazonka.Account

Description

Derived from API version 2021-02-01 of the AWS service descriptions, licensed under Apache 2.0.

Operations for Amazon Web Services Account Management

Service Configuration

defaultService :: Service

API version 2021-02-01 of the Amazon Account SDK configuration.

Errors

Error matchers are designed for use with the functions provided by Control.Exception.Lens. This allows catching (and rethrowing) service specific errors returned by Account.

ValidationException

_ValidationException :: AsError a => Getting (First ServiceError) a ServiceError

The operation failed because one of the input parameters was invalid.

AccessDeniedException

_AccessDeniedException :: AsError a => Getting (First ServiceError) a ServiceError

The operation failed because the calling identity doesn't have the minimum required permissions.

TooManyRequestsException

_TooManyRequestsException :: AsError a => Getting (First ServiceError) a ServiceError

The operation failed because it was called too frequently and exceeded a throttle limit.

InternalServerException

_InternalServerException :: AsError a => Getting (First ServiceError) a ServiceError

The operation failed because of an error internal to Amazon Web Services. Try your operation again later.

ResourceNotFoundException

_ResourceNotFoundException :: AsError a => Getting (First ServiceError) a ServiceError

The operation failed because it specified a resource that can't be found.

Waiters

Waiters poll by repeatedly sending a request until some remote success condition configured by the Wait specification is fulfilled. The Wait specification determines how many attempts should be made, in addition to delay and retry strategies.

Operations

Some AWS operations return results that are incomplete and require subsequent requests in order to obtain the entire result set. The process of sending subsequent requests to continue where a previous request left off is called pagination. For example, the ListObjects operation of Amazon S3 returns up to 1000 objects at a time, and you must send subsequent requests with the appropriate Marker in order to retrieve the next page of results.

Operations that have an AWSPager instance can transparently perform subsequent requests, correctly setting Markers and other request facets to iterate through the entire result set of a truncated API operation. Operations which support this have an additional note in the documentation.

Many operations have the ability to filter results on the server side. See the individual operation parameters for details.

DeleteAlternateContact

data DeleteAlternateContact

See: newDeleteAlternateContact smart constructor.

Constructors

DeleteAlternateContact' (Maybe Text) AlternateContactType

newDeleteAlternateContact

Arguments

:: AlternateContactType
$sel:alternateContactType:DeleteAlternateContact'

-> DeleteAlternateContact

Create a value of DeleteAlternateContact with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:accountId:DeleteAlternateContact', deleteAlternateContact_accountId - Specifies the 12 digit account ID number of the Amazon Web Services account that you want to access or modify with this operation.

If you do not specify this parameter, it defaults to the Amazon Web Services account of the identity used to call the operation.

To use this parameter, the caller must be an identity in the organization's management account or a delegated administrator account, and the specified account ID must be a member account in the same organization. The organization must have all features enabled, and the organization must have trusted access enabled for the Account Management service, and optionally a delegated admin account assigned.

The management account can't specify its own AccountId; it must call the operation in standalone context by not including the AccountId parameter.

To call this operation on an account that is not a member of an organization, then don't specify this parameter, and call the operation using an identity belonging to the account whose contacts you wish to retrieve or modify.

$sel:alternateContactType:DeleteAlternateContact', deleteAlternateContact_alternateContactType - Specifies which of the alternate contacts to delete.

data DeleteAlternateContactResponse

See: newDeleteAlternateContactResponse smart constructor.

Constructors

DeleteAlternateContactResponse'

newDeleteAlternateContactResponse :: DeleteAlternateContactResponse

Create a value of DeleteAlternateContactResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

PutAlternateContact

data PutAlternateContact

See: newPutAlternateContact smart constructor.

Constructors

PutAlternateContact' (Maybe Text) AlternateContactType (Sensitive Text) (Sensitive Text) (Sensitive Text) (Sensitive Text)

newPutAlternateContact

Arguments

:: AlternateContactType
$sel:alternateContactType:PutAlternateContact'

-> Text
$sel:emailAddress:PutAlternateContact'

-> Text
$sel:name:PutAlternateContact'

-> Text
$sel:phoneNumber:PutAlternateContact'

-> Text
$sel:title:PutAlternateContact'

-> PutAlternateContact

Create a value of PutAlternateContact with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:accountId:PutAlternateContact', putAlternateContact_accountId - Specifies the 12 digit account ID number of the Amazon Web Services account that you want to access or modify with this operation.

If you do not specify this parameter, it defaults to the Amazon Web Services account of the identity used to call the operation.

To use this parameter, the caller must be an identity in the organization's management account or a delegated administrator account, and the specified account ID must be a member account in the same organization. The organization must have all features enabled, and the organization must have trusted access enabled for the Account Management service, and optionally a delegated admin account assigned.

The management account can't specify its own AccountId; it must call the operation in standalone context by not including the AccountId parameter.

To call this operation on an account that is not a member of an organization, then don't specify this parameter, and call the operation using an identity belonging to the account whose contacts you wish to retrieve or modify.

$sel:alternateContactType:PutAlternateContact', putAlternateContact_alternateContactType - Specifies which alternate contact you want to create or update.

$sel:emailAddress:PutAlternateContact', putAlternateContact_emailAddress - Specifies an email address for the alternate contact.

$sel:name:PutAlternateContact', putAlternateContact_name - Specifies a name for the alternate contact.

$sel:phoneNumber:PutAlternateContact', putAlternateContact_phoneNumber - Specifies a phone number for the alternate contact.

$sel:title:PutAlternateContact', putAlternateContact_title - Specifies a title for the alternate contact.

data PutAlternateContactResponse

See: newPutAlternateContactResponse smart constructor.

Constructors

PutAlternateContactResponse'

newPutAlternateContactResponse :: PutAlternateContactResponse

Create a value of PutAlternateContactResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

GetAlternateContact

data GetAlternateContact

See: newGetAlternateContact smart constructor.

Constructors

GetAlternateContact' (Maybe Text) AlternateContactType

newGetAlternateContact

Arguments

:: AlternateContactType
$sel:alternateContactType:GetAlternateContact'

-> GetAlternateContact

Create a value of GetAlternateContact with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:accountId:GetAlternateContact', getAlternateContact_accountId - Specifies the 12 digit account ID number of the Amazon Web Services account that you want to access or modify with this operation.

If you do not specify this parameter, it defaults to the Amazon Web Services account of the identity used to call the operation.

To use this parameter, the caller must be an identity in the organization's management account or a delegated administrator account, and the specified account ID must be a member account in the same organization. The organization must have all features enabled, and the organization must have trusted access enabled for the Account Management service, and optionally a delegated admin account assigned.

The management account can't specify its own AccountId; it must call the operation in standalone context by not including the AccountId parameter.

To call this operation on an account that is not a member of an organization, then don't specify this parameter, and call the operation using an identity belonging to the account whose contacts you wish to retrieve or modify.

$sel:alternateContactType:GetAlternateContact', getAlternateContact_alternateContactType - Specifies which alternate contact you want to retrieve.

data GetAlternateContactResponse

See: newGetAlternateContactResponse smart constructor.

Constructors

GetAlternateContactResponse' (Maybe AlternateContact) Int

newGetAlternateContactResponse

Arguments

:: Int
$sel:httpStatus:GetAlternateContactResponse'

-> GetAlternateContactResponse

Create a value of GetAlternateContactResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:alternateContact:GetAlternateContactResponse', getAlternateContactResponse_alternateContact - A structure that contains the details for the specified alternate contact.

$sel:httpStatus:GetAlternateContactResponse', getAlternateContactResponse_httpStatus - The response's http status code.

Types

AlternateContactType

newtype AlternateContactType

Constructors

AlternateContactType'

Fields

fromAlternateContactType :: Text

Bundled Patterns

pattern AlternateContactType_BILLING :: AlternateContactType

pattern AlternateContactType_OPERATIONS :: AlternateContactType

pattern AlternateContactType_SECURITY :: AlternateContactType

AlternateContact

data AlternateContact

A structure that contains the details of an alternate contact associated with an Amazon Web Services account

See: newAlternateContact smart constructor.

Constructors

AlternateContact' (Maybe AlternateContactType) (Maybe (Sensitive Text)) (Maybe (Sensitive Text)) (Maybe (Sensitive Text)) (Maybe (Sensitive Text))

newAlternateContact :: AlternateContact

Create a value of AlternateContact with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:alternateContactType:AlternateContact', alternateContact_alternateContactType - The type of alternate contact.

$sel:phoneNumber:AlternateContact', alternateContact_phoneNumber - The phone number associated with this alternate contact.

$sel:name:AlternateContact', alternateContact_name - The name associated with this alternate contact.

$sel:emailAddress:AlternateContact', alternateContact_emailAddress - The email address associated with this alternate contact.

$sel:title:AlternateContact', alternateContact_title - The title associated with this alternate contact.