Amazonka.Account.Lens

Description

Operations

DeleteAlternateContact

deleteAlternateContact_accountId :: Lens' DeleteAlternateContact (Maybe Text)

Specifies the 12 digit account ID number of the Amazon Web Services account that you want to access or modify with this operation.

If you do not specify this parameter, it defaults to the Amazon Web Services account of the identity used to call the operation.

To use this parameter, the caller must be an identity in the organization's management account or a delegated administrator account, and the specified account ID must be a member account in the same organization. The organization must have all features enabled, and the organization must have trusted access enabled for the Account Management service, and optionally a delegated admin account assigned.

The management account can't specify its own AccountId; it must call the operation in standalone context by not including the AccountId parameter.

To call this operation on an account that is not a member of an organization, then don't specify this parameter, and call the operation using an identity belonging to the account whose contacts you wish to retrieve or modify.

deleteAlternateContact_alternateContactType :: Lens' DeleteAlternateContact AlternateContactType

Specifies which of the alternate contacts to delete.

PutAlternateContact

putAlternateContact_accountId :: Lens' PutAlternateContact (Maybe Text)

Specifies the 12 digit account ID number of the Amazon Web Services account that you want to access or modify with this operation.

If you do not specify this parameter, it defaults to the Amazon Web Services account of the identity used to call the operation.

To use this parameter, the caller must be an identity in the organization's management account or a delegated administrator account, and the specified account ID must be a member account in the same organization. The organization must have all features enabled, and the organization must have trusted access enabled for the Account Management service, and optionally a delegated admin account assigned.

The management account can't specify its own AccountId; it must call the operation in standalone context by not including the AccountId parameter.

To call this operation on an account that is not a member of an organization, then don't specify this parameter, and call the operation using an identity belonging to the account whose contacts you wish to retrieve or modify.

putAlternateContact_alternateContactType :: Lens' PutAlternateContact AlternateContactType

Specifies which alternate contact you want to create or update.

putAlternateContact_emailAddress :: Lens' PutAlternateContact Text

Specifies an email address for the alternate contact.

putAlternateContact_name :: Lens' PutAlternateContact Text

Specifies a name for the alternate contact.

putAlternateContact_phoneNumber :: Lens' PutAlternateContact Text

Specifies a phone number for the alternate contact.

putAlternateContact_title :: Lens' PutAlternateContact Text

Specifies a title for the alternate contact.

GetAlternateContact

getAlternateContact_accountId :: Lens' GetAlternateContact (Maybe Text)

Specifies the 12 digit account ID number of the Amazon Web Services account that you want to access or modify with this operation.

If you do not specify this parameter, it defaults to the Amazon Web Services account of the identity used to call the operation.

To use this parameter, the caller must be an identity in the organization's management account or a delegated administrator account, and the specified account ID must be a member account in the same organization. The organization must have all features enabled, and the organization must have trusted access enabled for the Account Management service, and optionally a delegated admin account assigned.

The management account can't specify its own AccountId; it must call the operation in standalone context by not including the AccountId parameter.

To call this operation on an account that is not a member of an organization, then don't specify this parameter, and call the operation using an identity belonging to the account whose contacts you wish to retrieve or modify.

getAlternateContact_alternateContactType :: Lens' GetAlternateContact AlternateContactType

Specifies which alternate contact you want to retrieve.

getAlternateContactResponse_alternateContact :: Lens' GetAlternateContactResponse (Maybe AlternateContact)

A structure that contains the details for the specified alternate contact.

getAlternateContactResponse_httpStatus :: Lens' GetAlternateContactResponse Int

The response's http status code.

Types

AlternateContact

alternateContact_alternateContactType :: Lens' AlternateContact (Maybe AlternateContactType)

The type of alternate contact.

alternateContact_phoneNumber :: Lens' AlternateContact (Maybe Text)

The phone number associated with this alternate contact.

alternateContact_name :: Lens' AlternateContact (Maybe Text)

The name associated with this alternate contact.

alternateContact_emailAddress :: Lens' AlternateContact (Maybe Text)

The email address associated with this alternate contact.

alternateContact_title :: Lens' AlternateContact (Maybe Text)

The title associated with this alternate contact.