Amazonka.Account.Types.AlternateContact

Description

Documentation

data AlternateContact

A structure that contains the details of an alternate contact associated with an Amazon Web Services account

See: newAlternateContact smart constructor.

Constructors

AlternateContact'

Fields

alternateContactType :: Maybe AlternateContactType

The type of alternate contact.

phoneNumber :: Maybe (Sensitive Text)

The phone number associated with this alternate contact.

name :: Maybe (Sensitive Text)

The name associated with this alternate contact.

emailAddress :: Maybe (Sensitive Text)

The email address associated with this alternate contact.

title :: Maybe (Sensitive Text)

The title associated with this alternate contact.

newAlternateContact :: AlternateContact

Create a value of AlternateContact with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:alternateContactType:AlternateContact', alternateContact_alternateContactType - The type of alternate contact.

$sel:phoneNumber:AlternateContact', alternateContact_phoneNumber - The phone number associated with this alternate contact.

$sel:name:AlternateContact', alternateContact_name - The name associated with this alternate contact.

$sel:emailAddress:AlternateContact', alternateContact_emailAddress - The email address associated with this alternate contact.

$sel:title:AlternateContact', alternateContact_title - The title associated with this alternate contact.

alternateContact_alternateContactType :: Lens' AlternateContact (Maybe AlternateContactType)

The type of alternate contact.

alternateContact_phoneNumber :: Lens' AlternateContact (Maybe Text)

The phone number associated with this alternate contact.

alternateContact_name :: Lens' AlternateContact (Maybe Text)

The name associated with this alternate contact.

alternateContact_emailAddress :: Lens' AlternateContact (Maybe Text)

The email address associated with this alternate contact.

alternateContact_title :: Lens' AlternateContact (Maybe Text)

The title associated with this alternate contact.