Amazonka.AlexaBusiness.CreateGatewayGroup

Description

Creates a gateway group with the specified details.

Creating a Request

data CreateGatewayGroup

See: newCreateGatewayGroup smart constructor.

Constructors

CreateGatewayGroup'

Fields

description :: Maybe Text

The description of the gateway group.

tags :: Maybe [Tag]

The tags to be added to the specified resource. Do not provide system tags.

name :: Text

The name of the gateway group.

clientRequestToken :: Text

A unique, user-specified identifier for the request that ensures idempotency.

newCreateGatewayGroup

Arguments

:: Text
$sel:name:CreateGatewayGroup'

-> Text
$sel:clientRequestToken:CreateGatewayGroup'

-> CreateGatewayGroup

Create a value of CreateGatewayGroup with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:description:CreateGatewayGroup', createGatewayGroup_description - The description of the gateway group.

$sel:tags:CreateGatewayGroup', createGatewayGroup_tags - The tags to be added to the specified resource. Do not provide system tags.

$sel:name:CreateGatewayGroup', createGatewayGroup_name - The name of the gateway group.

$sel:clientRequestToken:CreateGatewayGroup', createGatewayGroup_clientRequestToken - A unique, user-specified identifier for the request that ensures idempotency.

Request Lenses

createGatewayGroup_description :: Lens' CreateGatewayGroup (Maybe Text)

The description of the gateway group.

createGatewayGroup_tags :: Lens' CreateGatewayGroup (Maybe [Tag])

The tags to be added to the specified resource. Do not provide system tags.

createGatewayGroup_name :: Lens' CreateGatewayGroup Text

The name of the gateway group.

createGatewayGroup_clientRequestToken :: Lens' CreateGatewayGroup Text

A unique, user-specified identifier for the request that ensures idempotency.

Destructuring the Response

data CreateGatewayGroupResponse

See: newCreateGatewayGroupResponse smart constructor.

Constructors

CreateGatewayGroupResponse'

Fields

gatewayGroupArn :: Maybe Text

The ARN of the created gateway group.

httpStatus :: Int

The response's http status code.

newCreateGatewayGroupResponse

Arguments

:: Int
$sel:httpStatus:CreateGatewayGroupResponse'

-> CreateGatewayGroupResponse

Create a value of CreateGatewayGroupResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:gatewayGroupArn:CreateGatewayGroupResponse', createGatewayGroupResponse_gatewayGroupArn - The ARN of the created gateway group.

$sel:httpStatus:CreateGatewayGroupResponse', createGatewayGroupResponse_httpStatus - The response's http status code.

Response Lenses

createGatewayGroupResponse_gatewayGroupArn :: Lens' CreateGatewayGroupResponse (Maybe Text)

The ARN of the created gateway group.

createGatewayGroupResponse_httpStatus :: Lens' CreateGatewayGroupResponse Int

The response's http status code.