Amazonka.AlexaBusiness.ForgetSmartHomeAppliances

Description

Forgets smart home appliances associated to a room.

Creating a Request

data ForgetSmartHomeAppliances

See: newForgetSmartHomeAppliances smart constructor.

Constructors

ForgetSmartHomeAppliances'

Fields

roomArn :: Text

The room that the appliances are associated with.

newForgetSmartHomeAppliances

Arguments

:: Text
$sel:roomArn:ForgetSmartHomeAppliances'

-> ForgetSmartHomeAppliances

Create a value of ForgetSmartHomeAppliances with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:roomArn:ForgetSmartHomeAppliances', forgetSmartHomeAppliances_roomArn - The room that the appliances are associated with.

Request Lenses

forgetSmartHomeAppliances_roomArn :: Lens' ForgetSmartHomeAppliances Text

The room that the appliances are associated with.

Destructuring the Response

data ForgetSmartHomeAppliancesResponse

See: newForgetSmartHomeAppliancesResponse smart constructor.

Constructors

ForgetSmartHomeAppliancesResponse'

Fields

httpStatus :: Int

The response's http status code.

newForgetSmartHomeAppliancesResponse

Arguments

:: Int
$sel:httpStatus:ForgetSmartHomeAppliancesResponse'

-> ForgetSmartHomeAppliancesResponse

Create a value of ForgetSmartHomeAppliancesResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:httpStatus:ForgetSmartHomeAppliancesResponse', forgetSmartHomeAppliancesResponse_httpStatus - The response's http status code.

Response Lenses

forgetSmartHomeAppliancesResponse_httpStatus :: Lens' ForgetSmartHomeAppliancesResponse Int

The response's http status code.