Amazonka.AlexaBusiness.GetRoomSkillParameter

Description

Gets room skill parameter details by room, skill, and parameter key ARN.

Creating a Request

data GetRoomSkillParameter

See: newGetRoomSkillParameter smart constructor.

Constructors

GetRoomSkillParameter'

Fields

roomArn :: Maybe Text

The ARN of the room from which to get the room skill parameter details.

skillId :: Text

The ARN of the skill from which to get the room skill parameter details. Required.

parameterKey :: Text

The room skill parameter key for which to get details. Required.

newGetRoomSkillParameter

Arguments

:: Text
$sel:skillId:GetRoomSkillParameter'

-> Text
$sel:parameterKey:GetRoomSkillParameter'

-> GetRoomSkillParameter

Create a value of GetRoomSkillParameter with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:roomArn:GetRoomSkillParameter', getRoomSkillParameter_roomArn - The ARN of the room from which to get the room skill parameter details.

$sel:skillId:GetRoomSkillParameter', getRoomSkillParameter_skillId - The ARN of the skill from which to get the room skill parameter details. Required.

$sel:parameterKey:GetRoomSkillParameter', getRoomSkillParameter_parameterKey - The room skill parameter key for which to get details. Required.

Request Lenses

getRoomSkillParameter_roomArn :: Lens' GetRoomSkillParameter (Maybe Text)

The ARN of the room from which to get the room skill parameter details.

getRoomSkillParameter_skillId :: Lens' GetRoomSkillParameter Text

The ARN of the skill from which to get the room skill parameter details. Required.

getRoomSkillParameter_parameterKey :: Lens' GetRoomSkillParameter Text

The room skill parameter key for which to get details. Required.

Destructuring the Response

data GetRoomSkillParameterResponse

See: newGetRoomSkillParameterResponse smart constructor.

Constructors

GetRoomSkillParameterResponse'

Fields

roomSkillParameter :: Maybe RoomSkillParameter

The details of the room skill parameter requested. Required.

httpStatus :: Int

The response's http status code.

newGetRoomSkillParameterResponse

Arguments

:: Int
$sel:httpStatus:GetRoomSkillParameterResponse'

-> GetRoomSkillParameterResponse

Create a value of GetRoomSkillParameterResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:roomSkillParameter:GetRoomSkillParameterResponse', getRoomSkillParameterResponse_roomSkillParameter - The details of the room skill parameter requested. Required.

$sel:httpStatus:GetRoomSkillParameterResponse', getRoomSkillParameterResponse_httpStatus - The response's http status code.

Response Lenses

getRoomSkillParameterResponse_roomSkillParameter :: Lens' GetRoomSkillParameterResponse (Maybe RoomSkillParameter)

The details of the room skill parameter requested. Required.

getRoomSkillParameterResponse_httpStatus :: Lens' GetRoomSkillParameterResponse Int

The response's http status code.