Amazonka.AlexaBusiness.SearchUsers

Description

Searches users and lists the ones that meet a set of filter and sort criteria.

This operation returns paginated results.

Creating a Request

data SearchUsers

See: newSearchUsers smart constructor.

Constructors

SearchUsers'

Fields

filters :: Maybe [Filter]

The filters to use for listing a specific set of users. Required. Supported filter keys are UserId, FirstName, LastName, Email, and EnrollmentStatus.

sortCriteria :: Maybe [Sort]

The sort order to use in listing the filtered set of users. Required. Supported sort keys are UserId, FirstName, LastName, Email, and EnrollmentStatus.

nextToken :: Maybe Text

An optional token returned from a prior request. Use this token for pagination of results from this action. If this parameter is specified, the response includes only results beyond the token, up to the value specified by MaxResults. Required.

maxResults :: Maybe Natural

The maximum number of results to include in the response. If more results exist than the specified MaxResults value, a token is included in the response so that the remaining results can be retrieved. Required.

newSearchUsers :: SearchUsers

Create a value of SearchUsers with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:filters:SearchUsers', searchUsers_filters - The filters to use for listing a specific set of users. Required. Supported filter keys are UserId, FirstName, LastName, Email, and EnrollmentStatus.

$sel:sortCriteria:SearchUsers', searchUsers_sortCriteria - The sort order to use in listing the filtered set of users. Required. Supported sort keys are UserId, FirstName, LastName, Email, and EnrollmentStatus.

$sel:nextToken:SearchUsers', searchUsers_nextToken - An optional token returned from a prior request. Use this token for pagination of results from this action. If this parameter is specified, the response includes only results beyond the token, up to the value specified by MaxResults. Required.

$sel:maxResults:SearchUsers', searchUsers_maxResults - The maximum number of results to include in the response. If more results exist than the specified MaxResults value, a token is included in the response so that the remaining results can be retrieved. Required.

Request Lenses

searchUsers_filters :: Lens' SearchUsers (Maybe [Filter])

The filters to use for listing a specific set of users. Required. Supported filter keys are UserId, FirstName, LastName, Email, and EnrollmentStatus.

searchUsers_sortCriteria :: Lens' SearchUsers (Maybe [Sort])

The sort order to use in listing the filtered set of users. Required. Supported sort keys are UserId, FirstName, LastName, Email, and EnrollmentStatus.

searchUsers_nextToken :: Lens' SearchUsers (Maybe Text)

An optional token returned from a prior request. Use this token for pagination of results from this action. If this parameter is specified, the response includes only results beyond the token, up to the value specified by MaxResults. Required.

searchUsers_maxResults :: Lens' SearchUsers (Maybe Natural)

The maximum number of results to include in the response. If more results exist than the specified MaxResults value, a token is included in the response so that the remaining results can be retrieved. Required.

Destructuring the Response

data SearchUsersResponse

See: newSearchUsersResponse smart constructor.

Constructors

SearchUsersResponse'

Fields

users :: Maybe [UserData]

The users that meet the specified set of filter criteria, in sort order.

nextToken :: Maybe Text

The token returned to indicate that there is more data available.

totalCount :: Maybe Int

The total number of users returned.

httpStatus :: Int

The response's http status code.

newSearchUsersResponse

Arguments

:: Int
$sel:httpStatus:SearchUsersResponse'

-> SearchUsersResponse

Create a value of SearchUsersResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:users:SearchUsersResponse', searchUsersResponse_users - The users that meet the specified set of filter criteria, in sort order.

$sel:nextToken:SearchUsers', searchUsersResponse_nextToken - The token returned to indicate that there is more data available.

$sel:totalCount:SearchUsersResponse', searchUsersResponse_totalCount - The total number of users returned.

$sel:httpStatus:SearchUsersResponse', searchUsersResponse_httpStatus - The response's http status code.

Response Lenses

searchUsersResponse_users :: Lens' SearchUsersResponse (Maybe [UserData])

The users that meet the specified set of filter criteria, in sort order.

searchUsersResponse_nextToken :: Lens' SearchUsersResponse (Maybe Text)

The token returned to indicate that there is more data available.

searchUsersResponse_totalCount :: Lens' SearchUsersResponse (Maybe Int)

The total number of users returned.

searchUsersResponse_httpStatus :: Lens' SearchUsersResponse Int

The response's http status code.