Amazonka.AlexaBusiness.Types.Contact

Description

Documentation

data Contact

A contact with attributes.

See: newContact smart constructor.

Constructors

Contact'

Fields

lastName :: Maybe Text

The last name of the contact, used to call the contact on the device.

contactArn :: Maybe Text

The ARN of the contact.

phoneNumbers :: Maybe [PhoneNumber]

The list of phone numbers for the contact.

phoneNumber :: Maybe (Sensitive Text)

The phone number of the contact. The phone number type defaults to WORK. You can either specify PhoneNumber or PhoneNumbers. We recommend that you use PhoneNumbers, which lets you specify the phone number type and multiple numbers.

sipAddresses :: Maybe [SipAddress]

The list of SIP addresses for the contact.

firstName :: Maybe Text

The first name of the contact, used to call the contact on the device.

displayName :: Maybe Text

The name of the contact to display on the console.

newContact :: Contact

Create a value of Contact with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:lastName:Contact', contact_lastName - The last name of the contact, used to call the contact on the device.

$sel:contactArn:Contact', contact_contactArn - The ARN of the contact.

$sel:phoneNumbers:Contact', contact_phoneNumbers - The list of phone numbers for the contact.

$sel:phoneNumber:Contact', contact_phoneNumber - The phone number of the contact. The phone number type defaults to WORK. You can either specify PhoneNumber or PhoneNumbers. We recommend that you use PhoneNumbers, which lets you specify the phone number type and multiple numbers.

$sel:sipAddresses:Contact', contact_sipAddresses - The list of SIP addresses for the contact.

$sel:firstName:Contact', contact_firstName - The first name of the contact, used to call the contact on the device.

$sel:displayName:Contact', contact_displayName - The name of the contact to display on the console.

contact_lastName :: Lens' Contact (Maybe Text)

The last name of the contact, used to call the contact on the device.

contact_contactArn :: Lens' Contact (Maybe Text)

The ARN of the contact.

contact_phoneNumbers :: Lens' Contact (Maybe [PhoneNumber])

The list of phone numbers for the contact.

contact_phoneNumber :: Lens' Contact (Maybe Text)

The phone number of the contact. The phone number type defaults to WORK. You can either specify PhoneNumber or PhoneNumbers. We recommend that you use PhoneNumbers, which lets you specify the phone number type and multiple numbers.

contact_sipAddresses :: Lens' Contact (Maybe [SipAddress])

The list of SIP addresses for the contact.

contact_firstName :: Lens' Contact (Maybe Text)

The first name of the contact, used to call the contact on the device.

contact_displayName :: Lens' Contact (Maybe Text)

The name of the contact to display on the console.