Amazonka.AlexaBusiness.Types.GatewaySummary

Description

Documentation

data GatewaySummary

The summary of a gateway.

See: newGatewaySummary smart constructor.

Constructors

GatewaySummary'

Fields

arn :: Maybe Text

The ARN of the gateway.

name :: Maybe Text

The name of the gateway.

gatewayGroupArn :: Maybe Text

The ARN of the gateway group that the gateway is associated to.

softwareVersion :: Maybe Text

The software version of the gateway. The gateway automatically updates its software version during normal operation.

description :: Maybe Text

The description of the gateway.

newGatewaySummary :: GatewaySummary

Create a value of GatewaySummary with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:arn:GatewaySummary', gatewaySummary_arn - The ARN of the gateway.

$sel:name:GatewaySummary', gatewaySummary_name - The name of the gateway.

$sel:gatewayGroupArn:GatewaySummary', gatewaySummary_gatewayGroupArn - The ARN of the gateway group that the gateway is associated to.

$sel:softwareVersion:GatewaySummary', gatewaySummary_softwareVersion - The software version of the gateway. The gateway automatically updates its software version during normal operation.

$sel:description:GatewaySummary', gatewaySummary_description - The description of the gateway.

gatewaySummary_arn :: Lens' GatewaySummary (Maybe Text)

The ARN of the gateway.

gatewaySummary_name :: Lens' GatewaySummary (Maybe Text)

The name of the gateway.

gatewaySummary_gatewayGroupArn :: Lens' GatewaySummary (Maybe Text)

The ARN of the gateway group that the gateway is associated to.

gatewaySummary_softwareVersion :: Lens' GatewaySummary (Maybe Text)

The software version of the gateway. The gateway automatically updates its software version during normal operation.

gatewaySummary_description :: Lens' GatewaySummary (Maybe Text)

The description of the gateway.