Amazonka.AlexaBusiness.Types.MeetingRoomConfiguration

Description

Documentation

data MeetingRoomConfiguration

Meeting room settings of a room profile.

See: newMeetingRoomConfiguration smart constructor.

Constructors

MeetingRoomConfiguration'

Fields

instantBooking :: Maybe InstantBooking

Settings to automatically book the room if available for a configured duration when joining a meeting with Alexa.

endOfMeetingReminder :: Maybe EndOfMeetingReminder

Settings for the end of meeting reminder feature that are applied to a room profile. The end of meeting reminder enables Alexa to remind users when a meeting is ending.

requireCheckIn :: Maybe RequireCheckIn

Settings for requiring a check in when a room is reserved. Alexa can cancel a room reservation if it's not checked into. This makes the room available for others. Users can check in by joining the meeting with Alexa or an AVS device, or by saying “Alexa, check in.”

roomUtilizationMetricsEnabled :: Maybe Bool

Whether room utilization metrics are enabled or not.

newMeetingRoomConfiguration :: MeetingRoomConfiguration

Create a value of MeetingRoomConfiguration with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:instantBooking:MeetingRoomConfiguration', meetingRoomConfiguration_instantBooking - Settings to automatically book the room if available for a configured duration when joining a meeting with Alexa.

$sel:endOfMeetingReminder:MeetingRoomConfiguration', meetingRoomConfiguration_endOfMeetingReminder - Settings for the end of meeting reminder feature that are applied to a room profile. The end of meeting reminder enables Alexa to remind users when a meeting is ending.

$sel:requireCheckIn:MeetingRoomConfiguration', meetingRoomConfiguration_requireCheckIn - Settings for requiring a check in when a room is reserved. Alexa can cancel a room reservation if it's not checked into. This makes the room available for others. Users can check in by joining the meeting with Alexa or an AVS device, or by saying “Alexa, check in.”

$sel:roomUtilizationMetricsEnabled:MeetingRoomConfiguration', meetingRoomConfiguration_roomUtilizationMetricsEnabled - Whether room utilization metrics are enabled or not.

meetingRoomConfiguration_instantBooking :: Lens' MeetingRoomConfiguration (Maybe InstantBooking)

Settings to automatically book the room if available for a configured duration when joining a meeting with Alexa.

meetingRoomConfiguration_endOfMeetingReminder :: Lens' MeetingRoomConfiguration (Maybe EndOfMeetingReminder)

Settings for the end of meeting reminder feature that are applied to a room profile. The end of meeting reminder enables Alexa to remind users when a meeting is ending.

meetingRoomConfiguration_requireCheckIn :: Lens' MeetingRoomConfiguration (Maybe RequireCheckIn)

Settings for requiring a check in when a room is reserved. Alexa can cancel a room reservation if it's not checked into. This makes the room available for others. Users can check in by joining the meeting with Alexa or an AVS device, or by saying “Alexa, check in.”

meetingRoomConfiguration_roomUtilizationMetricsEnabled :: Lens' MeetingRoomConfiguration (Maybe Bool)

Whether room utilization metrics are enabled or not.