Amazonka.AlexaBusiness.Types.UpdateEndOfMeetingReminder

Description

Documentation

data UpdateEndOfMeetingReminder

Settings for the end of meeting reminder feature that are applied to a room profile. The end of meeting reminder enables Alexa to remind users when a meeting is ending.

See: newUpdateEndOfMeetingReminder smart constructor.

Constructors

UpdateEndOfMeetingReminder'

Fields

enabled :: Maybe Bool

Whether an end of meeting reminder is enabled or not.

reminderAtMinutes :: Maybe (NonEmpty Int)

Updates settings for the end of meeting reminder feature that are applied to a room profile. The end of meeting reminder enables Alexa to remind users when a meeting is ending.

reminderType :: Maybe EndOfMeetingReminderType

The type of sound that users hear during the end of meeting reminder.

newUpdateEndOfMeetingReminder :: UpdateEndOfMeetingReminder

Create a value of UpdateEndOfMeetingReminder with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:enabled:UpdateEndOfMeetingReminder', updateEndOfMeetingReminder_enabled - Whether an end of meeting reminder is enabled or not.

$sel:reminderAtMinutes:UpdateEndOfMeetingReminder', updateEndOfMeetingReminder_reminderAtMinutes - Updates settings for the end of meeting reminder feature that are applied to a room profile. The end of meeting reminder enables Alexa to remind users when a meeting is ending.

$sel:reminderType:UpdateEndOfMeetingReminder', updateEndOfMeetingReminder_reminderType - The type of sound that users hear during the end of meeting reminder.

updateEndOfMeetingReminder_enabled :: Lens' UpdateEndOfMeetingReminder (Maybe Bool)

Whether an end of meeting reminder is enabled or not.

updateEndOfMeetingReminder_reminderAtMinutes :: Lens' UpdateEndOfMeetingReminder (Maybe (NonEmpty Int))

Updates settings for the end of meeting reminder feature that are applied to a room profile. The end of meeting reminder enables Alexa to remind users when a meeting is ending.

updateEndOfMeetingReminder_reminderType :: Lens' UpdateEndOfMeetingReminder (Maybe EndOfMeetingReminderType)

The type of sound that users hear during the end of meeting reminder.