Amazonka.AlexaBusiness.UpdateBusinessReportSchedule

Description

Updates the configuration of the report delivery schedule with the specified schedule ARN.

Creating a Request

data UpdateBusinessReportSchedule

See: newUpdateBusinessReportSchedule smart constructor.

Constructors

UpdateBusinessReportSchedule'

Fields

s3KeyPrefix :: Maybe Text

The S3 key where the report is delivered.

format :: Maybe BusinessReportFormat

The format of the generated report (individual CSV files or zipped files of individual files).

recurrence :: Maybe BusinessReportRecurrence

The recurrence of the reports.

scheduleName :: Maybe Text

The name identifier of the schedule.

s3BucketName :: Maybe Text

The S3 location of the output reports.

scheduleArn :: Text

The ARN of the business report schedule.

newUpdateBusinessReportSchedule

Arguments

:: Text
$sel:scheduleArn:UpdateBusinessReportSchedule'

-> UpdateBusinessReportSchedule

Create a value of UpdateBusinessReportSchedule with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:s3KeyPrefix:UpdateBusinessReportSchedule', updateBusinessReportSchedule_s3KeyPrefix - The S3 key where the report is delivered.

$sel:format:UpdateBusinessReportSchedule', updateBusinessReportSchedule_format - The format of the generated report (individual CSV files or zipped files of individual files).

$sel:recurrence:UpdateBusinessReportSchedule', updateBusinessReportSchedule_recurrence - The recurrence of the reports.

$sel:scheduleName:UpdateBusinessReportSchedule', updateBusinessReportSchedule_scheduleName - The name identifier of the schedule.

$sel:s3BucketName:UpdateBusinessReportSchedule', updateBusinessReportSchedule_s3BucketName - The S3 location of the output reports.

$sel:scheduleArn:UpdateBusinessReportSchedule', updateBusinessReportSchedule_scheduleArn - The ARN of the business report schedule.

Request Lenses

updateBusinessReportSchedule_s3KeyPrefix :: Lens' UpdateBusinessReportSchedule (Maybe Text)

The S3 key where the report is delivered.

updateBusinessReportSchedule_format :: Lens' UpdateBusinessReportSchedule (Maybe BusinessReportFormat)

The format of the generated report (individual CSV files or zipped files of individual files).

updateBusinessReportSchedule_recurrence :: Lens' UpdateBusinessReportSchedule (Maybe BusinessReportRecurrence)

The recurrence of the reports.

updateBusinessReportSchedule_scheduleName :: Lens' UpdateBusinessReportSchedule (Maybe Text)

The name identifier of the schedule.

updateBusinessReportSchedule_s3BucketName :: Lens' UpdateBusinessReportSchedule (Maybe Text)

The S3 location of the output reports.

updateBusinessReportSchedule_scheduleArn :: Lens' UpdateBusinessReportSchedule Text

The ARN of the business report schedule.

Destructuring the Response

data UpdateBusinessReportScheduleResponse

See: newUpdateBusinessReportScheduleResponse smart constructor.

Constructors

UpdateBusinessReportScheduleResponse'

Fields

httpStatus :: Int

The response's http status code.

newUpdateBusinessReportScheduleResponse

Arguments

:: Int
$sel:httpStatus:UpdateBusinessReportScheduleResponse'

-> UpdateBusinessReportScheduleResponse

Create a value of UpdateBusinessReportScheduleResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:httpStatus:UpdateBusinessReportScheduleResponse', updateBusinessReportScheduleResponse_httpStatus - The response's http status code.

Response Lenses

updateBusinessReportScheduleResponse_httpStatus :: Lens' UpdateBusinessReportScheduleResponse Int

The response's http status code.