Amazonka.Amplify.CreateBranch

Description

Creates a new branch for an Amplify app.

Creating a Request

data CreateBranch

The request structure for the create branch request.

See: newCreateBranch smart constructor.

Constructors

CreateBranch'

Fields

framework :: Maybe Text

The framework for the branch.

ttl :: Maybe Text

The content Time To Live (TTL) for the website in seconds.

enableNotification :: Maybe Bool

Enables notifications for the branch.

stage :: Maybe Stage

Describes the current stage for the branch.

backendEnvironmentArn :: Maybe Text

The Amazon Resource Name (ARN) for a backend environment that is part of an Amplify app.

enablePullRequestPreview :: Maybe Bool

Enables pull request previews for this branch.

basicAuthCredentials :: Maybe (Sensitive Text)

The basic authorization credentials for the branch.

buildSpec :: Maybe Text

The build specification (build spec) for the branch.

enablePerformanceMode :: Maybe Bool

Enables performance mode for the branch.

Performance mode optimizes for faster hosting performance by keeping content cached at the edge for a longer interval. When performance mode is enabled, hosting configuration or code changes can take up to 10 minutes to roll out.

displayName :: Maybe Text

The display name for a branch. This is used as the default domain prefix.

environmentVariables :: Maybe (HashMap Text Text)

The environment variables for the branch.

enableAutoBuild :: Maybe Bool

Enables auto building for the branch.

enableBasicAuth :: Maybe Bool

Enables basic authorization for the branch.

pullRequestEnvironmentName :: Maybe Text

The Amplify environment name for the pull request.

description :: Maybe Text

The description for the branch.

tags :: Maybe (HashMap Text Text)

The tag for the branch.

appId :: Text

The unique ID for an Amplify app.

branchName :: Text

The name for the branch.

newCreateBranch

Arguments

:: Text
$sel:appId:CreateBranch'

-> Text
$sel:branchName:CreateBranch'

-> CreateBranch

Create a value of CreateBranch with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:framework:CreateBranch', createBranch_framework - The framework for the branch.

$sel:ttl:CreateBranch', createBranch_ttl - The content Time To Live (TTL) for the website in seconds.

$sel:enableNotification:CreateBranch', createBranch_enableNotification - Enables notifications for the branch.

$sel:stage:CreateBranch', createBranch_stage - Describes the current stage for the branch.

$sel:backendEnvironmentArn:CreateBranch', createBranch_backendEnvironmentArn - The Amazon Resource Name (ARN) for a backend environment that is part of an Amplify app.

$sel:enablePullRequestPreview:CreateBranch', createBranch_enablePullRequestPreview - Enables pull request previews for this branch.

$sel:basicAuthCredentials:CreateBranch', createBranch_basicAuthCredentials - The basic authorization credentials for the branch.

$sel:buildSpec:CreateBranch', createBranch_buildSpec - The build specification (build spec) for the branch.

$sel:enablePerformanceMode:CreateBranch', createBranch_enablePerformanceMode - Enables performance mode for the branch.

Performance mode optimizes for faster hosting performance by keeping content cached at the edge for a longer interval. When performance mode is enabled, hosting configuration or code changes can take up to 10 minutes to roll out.

$sel:displayName:CreateBranch', createBranch_displayName - The display name for a branch. This is used as the default domain prefix.

$sel:environmentVariables:CreateBranch', createBranch_environmentVariables - The environment variables for the branch.

$sel:enableAutoBuild:CreateBranch', createBranch_enableAutoBuild - Enables auto building for the branch.

$sel:enableBasicAuth:CreateBranch', createBranch_enableBasicAuth - Enables basic authorization for the branch.

$sel:pullRequestEnvironmentName:CreateBranch', createBranch_pullRequestEnvironmentName - The Amplify environment name for the pull request.

$sel:description:CreateBranch', createBranch_description - The description for the branch.

$sel:tags:CreateBranch', createBranch_tags - The tag for the branch.

$sel:appId:CreateBranch', createBranch_appId - The unique ID for an Amplify app.

$sel:branchName:CreateBranch', createBranch_branchName - The name for the branch.

Request Lenses

createBranch_framework :: Lens' CreateBranch (Maybe Text)

The framework for the branch.

createBranch_ttl :: Lens' CreateBranch (Maybe Text)

The content Time To Live (TTL) for the website in seconds.

createBranch_enableNotification :: Lens' CreateBranch (Maybe Bool)

Enables notifications for the branch.

createBranch_stage :: Lens' CreateBranch (Maybe Stage)

Describes the current stage for the branch.

createBranch_backendEnvironmentArn :: Lens' CreateBranch (Maybe Text)

The Amazon Resource Name (ARN) for a backend environment that is part of an Amplify app.

createBranch_enablePullRequestPreview :: Lens' CreateBranch (Maybe Bool)

Enables pull request previews for this branch.

createBranch_basicAuthCredentials :: Lens' CreateBranch (Maybe Text)

The basic authorization credentials for the branch.

createBranch_buildSpec :: Lens' CreateBranch (Maybe Text)

The build specification (build spec) for the branch.

createBranch_enablePerformanceMode :: Lens' CreateBranch (Maybe Bool)

Enables performance mode for the branch.

Performance mode optimizes for faster hosting performance by keeping content cached at the edge for a longer interval. When performance mode is enabled, hosting configuration or code changes can take up to 10 minutes to roll out.

createBranch_displayName :: Lens' CreateBranch (Maybe Text)

The display name for a branch. This is used as the default domain prefix.

createBranch_environmentVariables :: Lens' CreateBranch (Maybe (HashMap Text Text))

The environment variables for the branch.

createBranch_enableAutoBuild :: Lens' CreateBranch (Maybe Bool)

Enables auto building for the branch.

createBranch_enableBasicAuth :: Lens' CreateBranch (Maybe Bool)

Enables basic authorization for the branch.

createBranch_pullRequestEnvironmentName :: Lens' CreateBranch (Maybe Text)

The Amplify environment name for the pull request.

createBranch_description :: Lens' CreateBranch (Maybe Text)

The description for the branch.

createBranch_tags :: Lens' CreateBranch (Maybe (HashMap Text Text))

The tag for the branch.

createBranch_appId :: Lens' CreateBranch Text

The unique ID for an Amplify app.

createBranch_branchName :: Lens' CreateBranch Text

The name for the branch.

Destructuring the Response

data CreateBranchResponse

The result structure for create branch request.

See: newCreateBranchResponse smart constructor.

Constructors

CreateBranchResponse'

Fields

httpStatus :: Int

The response's http status code.

branch :: Branch

Describes the branch for an Amplify app, which maps to a third-party repository branch.

newCreateBranchResponse

Arguments

:: Int
$sel:httpStatus:CreateBranchResponse'

-> Branch
$sel:branch:CreateBranchResponse'

-> CreateBranchResponse

Create a value of CreateBranchResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:httpStatus:CreateBranchResponse', createBranchResponse_httpStatus - The response's http status code.

$sel:branch:CreateBranchResponse', createBranchResponse_branch - Describes the branch for an Amplify app, which maps to a third-party repository branch.

Response Lenses

createBranchResponse_httpStatus :: Lens' CreateBranchResponse Int

The response's http status code.

createBranchResponse_branch :: Lens' CreateBranchResponse Branch

Describes the branch for an Amplify app, which maps to a third-party repository branch.