Amazonka.Amplify.CreateDeployment

Description

Creates a deployment for a manually deployed Amplify app. Manually deployed apps are not connected to a repository.

Creating a Request

data CreateDeployment

The request structure for the create a new deployment request.

See: newCreateDeployment smart constructor.

Constructors

CreateDeployment'

Fields

fileMap :: Maybe (HashMap Text Text)

An optional file map that contains the file name as the key and the file content md5 hash as the value. If this argument is provided, the service will generate a unique upload URL per file. Otherwise, the service will only generate a single upload URL for the zipped files.

appId :: Text

The unique ID for an Amplify app.

branchName :: Text

The name for the branch, for the job.

newCreateDeployment

Arguments

:: Text
$sel:appId:CreateDeployment'

-> Text
$sel:branchName:CreateDeployment'

-> CreateDeployment

Create a value of CreateDeployment with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:fileMap:CreateDeployment', createDeployment_fileMap - An optional file map that contains the file name as the key and the file content md5 hash as the value. If this argument is provided, the service will generate a unique upload URL per file. Otherwise, the service will only generate a single upload URL for the zipped files.

$sel:appId:CreateDeployment', createDeployment_appId - The unique ID for an Amplify app.

$sel:branchName:CreateDeployment', createDeployment_branchName - The name for the branch, for the job.

Request Lenses

createDeployment_fileMap :: Lens' CreateDeployment (Maybe (HashMap Text Text))

An optional file map that contains the file name as the key and the file content md5 hash as the value. If this argument is provided, the service will generate a unique upload URL per file. Otherwise, the service will only generate a single upload URL for the zipped files.

createDeployment_appId :: Lens' CreateDeployment Text

The unique ID for an Amplify app.

createDeployment_branchName :: Lens' CreateDeployment Text

The name for the branch, for the job.

Destructuring the Response

data CreateDeploymentResponse

The result structure for the create a new deployment request.

See: newCreateDeploymentResponse smart constructor.

Constructors

CreateDeploymentResponse'

Fields

jobId :: Maybe Text

The job ID for this deployment. will supply to start deployment api.

httpStatus :: Int

The response's http status code.

fileUploadUrls :: HashMap Text Text

When the fileMap argument is provided in the request, fileUploadUrls will contain a map of file names to upload URLs.

zipUploadUrl :: Text

When the fileMap argument is not provided in the request, this zipUploadUrl is returned.

newCreateDeploymentResponse

Arguments

:: Int
$sel:httpStatus:CreateDeploymentResponse'

-> Text
$sel:zipUploadUrl:CreateDeploymentResponse'

-> CreateDeploymentResponse

Create a value of CreateDeploymentResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:jobId:CreateDeploymentResponse', createDeploymentResponse_jobId - The job ID for this deployment. will supply to start deployment api.

$sel:httpStatus:CreateDeploymentResponse', createDeploymentResponse_httpStatus - The response's http status code.

$sel:fileUploadUrls:CreateDeploymentResponse', createDeploymentResponse_fileUploadUrls - When the fileMap argument is provided in the request, fileUploadUrls will contain a map of file names to upload URLs.

$sel:zipUploadUrl:CreateDeploymentResponse', createDeploymentResponse_zipUploadUrl - When the fileMap argument is not provided in the request, this zipUploadUrl is returned.

Response Lenses

createDeploymentResponse_jobId :: Lens' CreateDeploymentResponse (Maybe Text)

The job ID for this deployment. will supply to start deployment api.

createDeploymentResponse_httpStatus :: Lens' CreateDeploymentResponse Int

The response's http status code.

createDeploymentResponse_fileUploadUrls :: Lens' CreateDeploymentResponse (HashMap Text Text)

When the fileMap argument is provided in the request, fileUploadUrls will contain a map of file names to upload URLs.

createDeploymentResponse_zipUploadUrl :: Lens' CreateDeploymentResponse Text

When the fileMap argument is not provided in the request, this zipUploadUrl is returned.