Amazonka.Amplify.CreateWebhook

Description

Creates a new webhook on an Amplify app.

Creating a Request

data CreateWebhook

The request structure for the create webhook request.

See: newCreateWebhook smart constructor.

Constructors

CreateWebhook'

Fields

description :: Maybe Text

The description for a webhook.

appId :: Text

The unique ID for an Amplify app.

branchName :: Text

The name for a branch that is part of an Amplify app.

newCreateWebhook

Arguments

:: Text
$sel:appId:CreateWebhook'

-> Text
$sel:branchName:CreateWebhook'

-> CreateWebhook

Create a value of CreateWebhook with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:description:CreateWebhook', createWebhook_description - The description for a webhook.

$sel:appId:CreateWebhook', createWebhook_appId - The unique ID for an Amplify app.

$sel:branchName:CreateWebhook', createWebhook_branchName - The name for a branch that is part of an Amplify app.

Request Lenses

createWebhook_description :: Lens' CreateWebhook (Maybe Text)

The description for a webhook.

createWebhook_appId :: Lens' CreateWebhook Text

The unique ID for an Amplify app.

createWebhook_branchName :: Lens' CreateWebhook Text

The name for a branch that is part of an Amplify app.

Destructuring the Response

data CreateWebhookResponse

The result structure for the create webhook request.

See: newCreateWebhookResponse smart constructor.

Constructors

CreateWebhookResponse'

Fields

httpStatus :: Int

The response's http status code.

webhook :: Webhook

Describes a webhook that connects repository events to an Amplify app.

newCreateWebhookResponse

Arguments

:: Int
$sel:httpStatus:CreateWebhookResponse'

-> Webhook
$sel:webhook:CreateWebhookResponse'

-> CreateWebhookResponse

Create a value of CreateWebhookResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:httpStatus:CreateWebhookResponse', createWebhookResponse_httpStatus - The response's http status code.

$sel:webhook:CreateWebhookResponse', createWebhookResponse_webhook - Describes a webhook that connects repository events to an Amplify app.

Response Lenses

createWebhookResponse_httpStatus :: Lens' CreateWebhookResponse Int

The response's http status code.

createWebhookResponse_webhook :: Lens' CreateWebhookResponse Webhook

Describes a webhook that connects repository events to an Amplify app.