Amazonka.Amplify.Lens

Description

Operations

GetDomainAssociation

getDomainAssociation_appId :: Lens' GetDomainAssociation Text

The unique id for an Amplify app.

getDomainAssociation_domainName :: Lens' GetDomainAssociation Text

The name of the domain.

getDomainAssociationResponse_httpStatus :: Lens' GetDomainAssociationResponse Int

The response's http status code.

getDomainAssociationResponse_domainAssociation :: Lens' GetDomainAssociationResponse DomainAssociation

Describes the structure of a domain association, which associates a custom domain with an Amplify app.

ListArtifacts

listArtifacts_nextToken :: Lens' ListArtifacts (Maybe Text)

A pagination token. Set to null to start listing artifacts from start. If a non-null pagination token is returned in a result, pass its value in here to list more artifacts.

listArtifacts_maxResults :: Lens' ListArtifacts (Maybe Natural)

The maximum number of records to list in a single response.

listArtifacts_appId :: Lens' ListArtifacts Text

The unique ID for an Amplify app.

listArtifacts_branchName :: Lens' ListArtifacts Text

The name of a branch that is part of an Amplify app.

listArtifacts_jobId :: Lens' ListArtifacts Text

The unique ID for a job.

listArtifactsResponse_nextToken :: Lens' ListArtifactsResponse (Maybe Text)

A pagination token. If a non-null pagination token is returned in a result, pass its value in another request to retrieve more entries.

listArtifactsResponse_httpStatus :: Lens' ListArtifactsResponse Int

The response's http status code.

listArtifactsResponse_artifacts :: Lens' ListArtifactsResponse [Artifact]

A list of artifacts.

StopJob

stopJob_appId :: Lens' StopJob Text

The unique ID for an Amplify app.

stopJob_branchName :: Lens' StopJob Text

The name for the branch, for the job.

stopJob_jobId :: Lens' StopJob Text

The unique id for the job.

stopJobResponse_httpStatus :: Lens' StopJobResponse Int

The response's http status code.

stopJobResponse_jobSummary :: Lens' StopJobResponse JobSummary

The summary for the job.

GetBackendEnvironment

getBackendEnvironment_appId :: Lens' GetBackendEnvironment Text

The unique id for an Amplify app.

getBackendEnvironment_environmentName :: Lens' GetBackendEnvironment Text

The name for the backend environment.

getBackendEnvironmentResponse_httpStatus :: Lens' GetBackendEnvironmentResponse Int

The response's http status code.

getBackendEnvironmentResponse_backendEnvironment :: Lens' GetBackendEnvironmentResponse BackendEnvironment

Describes the backend environment for an Amplify app.

CreateWebhook

createWebhook_description :: Lens' CreateWebhook (Maybe Text)

The description for a webhook.

createWebhook_appId :: Lens' CreateWebhook Text

The unique ID for an Amplify app.

createWebhook_branchName :: Lens' CreateWebhook Text

The name for a branch that is part of an Amplify app.

createWebhookResponse_httpStatus :: Lens' CreateWebhookResponse Int

The response's http status code.

createWebhookResponse_webhook :: Lens' CreateWebhookResponse Webhook

Describes a webhook that connects repository events to an Amplify app.

GetBranch

getBranch_appId :: Lens' GetBranch Text

The unique ID for an Amplify app.

getBranch_branchName :: Lens' GetBranch Text

The name for the branch.

getBranchResponse_httpStatus :: Lens' GetBranchResponse Int

The response's http status code.

getBranchResponse_branch :: Lens' GetBranchResponse Branch

Undocumented member.

ListTagsForResource

listTagsForResource_resourceArn :: Lens' ListTagsForResource Text

The Amazon Resource Name (ARN) to use to list tags.

listTagsForResourceResponse_tags :: Lens' ListTagsForResourceResponse (Maybe (HashMap Text Text))

A list of tags for the specified The Amazon Resource Name (ARN).

listTagsForResourceResponse_httpStatus :: Lens' ListTagsForResourceResponse Int

The response's http status code.

CreateDomainAssociation

createDomainAssociation_enableAutoSubDomain :: Lens' CreateDomainAssociation (Maybe Bool)

Enables the automated creation of subdomains for branches.

createDomainAssociation_autoSubDomainCreationPatterns :: Lens' CreateDomainAssociation (Maybe [Text])

Sets the branch patterns for automatic subdomain creation.

createDomainAssociation_autoSubDomainIAMRole :: Lens' CreateDomainAssociation (Maybe Text)

The required AWS Identity and Access Management (IAM) service role for the Amazon Resource Name (ARN) for automatically creating subdomains.

createDomainAssociation_appId :: Lens' CreateDomainAssociation Text

The unique ID for an Amplify app.

createDomainAssociation_domainName :: Lens' CreateDomainAssociation Text

The domain name for the domain association.

createDomainAssociation_subDomainSettings :: Lens' CreateDomainAssociation [SubDomainSetting]

The setting for the subdomain.

createDomainAssociationResponse_httpStatus :: Lens' CreateDomainAssociationResponse Int

The response's http status code.

createDomainAssociationResponse_domainAssociation :: Lens' CreateDomainAssociationResponse DomainAssociation

Describes the structure of a domain association, which associates a custom domain with an Amplify app.

GetWebhook

getWebhook_webhookId :: Lens' GetWebhook Text

The unique ID for a webhook.

getWebhookResponse_httpStatus :: Lens' GetWebhookResponse Int

The response's http status code.

getWebhookResponse_webhook :: Lens' GetWebhookResponse Webhook

Describes the structure of a webhook.

DeleteBranch

deleteBranch_appId :: Lens' DeleteBranch Text

The unique ID for an Amplify app.

deleteBranch_branchName :: Lens' DeleteBranch Text

The name for the branch.

deleteBranchResponse_httpStatus :: Lens' DeleteBranchResponse Int

The response's http status code.

deleteBranchResponse_branch :: Lens' DeleteBranchResponse Branch

The branch for an Amplify app, which maps to a third-party repository branch.

UpdateBranch

updateBranch_framework :: Lens' UpdateBranch (Maybe Text)

The framework for the branch.

updateBranch_ttl :: Lens' UpdateBranch (Maybe Text)

The content Time to Live (TTL) for the website in seconds.

updateBranch_enableNotification :: Lens' UpdateBranch (Maybe Bool)

Enables notifications for the branch.

updateBranch_stage :: Lens' UpdateBranch (Maybe Stage)

Describes the current stage for the branch.

updateBranch_backendEnvironmentArn :: Lens' UpdateBranch (Maybe Text)

The Amazon Resource Name (ARN) for a backend environment that is part of an Amplify app.

updateBranch_enablePullRequestPreview :: Lens' UpdateBranch (Maybe Bool)

Enables pull request previews for this branch.

updateBranch_basicAuthCredentials :: Lens' UpdateBranch (Maybe Text)

The basic authorization credentials for the branch.

updateBranch_buildSpec :: Lens' UpdateBranch (Maybe Text)

The build specification (build spec) for the branch.

updateBranch_enablePerformanceMode :: Lens' UpdateBranch (Maybe Bool)

Enables performance mode for the branch.

Performance mode optimizes for faster hosting performance by keeping content cached at the edge for a longer interval. When performance mode is enabled, hosting configuration or code changes can take up to 10 minutes to roll out.

updateBranch_displayName :: Lens' UpdateBranch (Maybe Text)

The display name for a branch. This is used as the default domain prefix.

updateBranch_environmentVariables :: Lens' UpdateBranch (Maybe (HashMap Text Text))

The environment variables for the branch.

updateBranch_enableAutoBuild :: Lens' UpdateBranch (Maybe Bool)

Enables auto building for the branch.

updateBranch_enableBasicAuth :: Lens' UpdateBranch (Maybe Bool)

Enables basic authorization for the branch.

updateBranch_pullRequestEnvironmentName :: Lens' UpdateBranch (Maybe Text)

The Amplify environment name for the pull request.

updateBranch_description :: Lens' UpdateBranch (Maybe Text)

The description for the branch.

updateBranch_appId :: Lens' UpdateBranch Text

The unique ID for an Amplify app.

updateBranch_branchName :: Lens' UpdateBranch Text

The name for the branch.

updateBranchResponse_httpStatus :: Lens' UpdateBranchResponse Int

The response's http status code.

updateBranchResponse_branch :: Lens' UpdateBranchResponse Branch

The branch for an Amplify app, which maps to a third-party repository branch.

CreateBackendEnvironment

createBackendEnvironment_deploymentArtifacts :: Lens' CreateBackendEnvironment (Maybe Text)

The name of deployment artifacts.

createBackendEnvironment_stackName :: Lens' CreateBackendEnvironment (Maybe Text)

The AWS CloudFormation stack name of a backend environment.

createBackendEnvironment_appId :: Lens' CreateBackendEnvironment Text

The unique ID for an Amplify app.

createBackendEnvironment_environmentName :: Lens' CreateBackendEnvironment Text

The name for the backend environment.

createBackendEnvironmentResponse_httpStatus :: Lens' CreateBackendEnvironmentResponse Int

The response's http status code.

createBackendEnvironmentResponse_backendEnvironment :: Lens' CreateBackendEnvironmentResponse BackendEnvironment

Describes the backend environment for an Amplify app.

CreateDeployment

createDeployment_fileMap :: Lens' CreateDeployment (Maybe (HashMap Text Text))

An optional file map that contains the file name as the key and the file content md5 hash as the value. If this argument is provided, the service will generate a unique upload URL per file. Otherwise, the service will only generate a single upload URL for the zipped files.

createDeployment_appId :: Lens' CreateDeployment Text

The unique ID for an Amplify app.

createDeployment_branchName :: Lens' CreateDeployment Text

The name for the branch, for the job.

createDeploymentResponse_jobId :: Lens' CreateDeploymentResponse (Maybe Text)

The job ID for this deployment. will supply to start deployment api.

createDeploymentResponse_httpStatus :: Lens' CreateDeploymentResponse Int

The response's http status code.

createDeploymentResponse_fileUploadUrls :: Lens' CreateDeploymentResponse (HashMap Text Text)

When the fileMap argument is provided in the request, fileUploadUrls will contain a map of file names to upload URLs.

createDeploymentResponse_zipUploadUrl :: Lens' CreateDeploymentResponse Text

When the fileMap argument is not provided in the request, this zipUploadUrl is returned.

CreateBranch

createBranch_framework :: Lens' CreateBranch (Maybe Text)

The framework for the branch.

createBranch_ttl :: Lens' CreateBranch (Maybe Text)

The content Time To Live (TTL) for the website in seconds.

createBranch_enableNotification :: Lens' CreateBranch (Maybe Bool)

Enables notifications for the branch.

createBranch_stage :: Lens' CreateBranch (Maybe Stage)

Describes the current stage for the branch.

createBranch_backendEnvironmentArn :: Lens' CreateBranch (Maybe Text)

The Amazon Resource Name (ARN) for a backend environment that is part of an Amplify app.

createBranch_enablePullRequestPreview :: Lens' CreateBranch (Maybe Bool)

Enables pull request previews for this branch.

createBranch_basicAuthCredentials :: Lens' CreateBranch (Maybe Text)

The basic authorization credentials for the branch.

createBranch_buildSpec :: Lens' CreateBranch (Maybe Text)

The build specification (build spec) for the branch.

createBranch_enablePerformanceMode :: Lens' CreateBranch (Maybe Bool)

Enables performance mode for the branch.

Performance mode optimizes for faster hosting performance by keeping content cached at the edge for a longer interval. When performance mode is enabled, hosting configuration or code changes can take up to 10 minutes to roll out.

createBranch_displayName :: Lens' CreateBranch (Maybe Text)

The display name for a branch. This is used as the default domain prefix.

createBranch_environmentVariables :: Lens' CreateBranch (Maybe (HashMap Text Text))

The environment variables for the branch.

createBranch_enableAutoBuild :: Lens' CreateBranch (Maybe Bool)

Enables auto building for the branch.

createBranch_enableBasicAuth :: Lens' CreateBranch (Maybe Bool)

Enables basic authorization for the branch.

createBranch_pullRequestEnvironmentName :: Lens' CreateBranch (Maybe Text)

The Amplify environment name for the pull request.

createBranch_description :: Lens' CreateBranch (Maybe Text)

The description for the branch.

createBranch_tags :: Lens' CreateBranch (Maybe (HashMap Text Text))

The tag for the branch.

createBranch_appId :: Lens' CreateBranch Text

The unique ID for an Amplify app.

createBranch_branchName :: Lens' CreateBranch Text

The name for the branch.

createBranchResponse_httpStatus :: Lens' CreateBranchResponse Int

The response's http status code.

createBranchResponse_branch :: Lens' CreateBranchResponse Branch

Describes the branch for an Amplify app, which maps to a third-party repository branch.

GenerateAccessLogs

generateAccessLogs_startTime :: Lens' GenerateAccessLogs (Maybe UTCTime)

The time at which the logs should start. The time range specified is inclusive of the start time.

generateAccessLogs_endTime :: Lens' GenerateAccessLogs (Maybe UTCTime)

The time at which the logs should end. The time range specified is inclusive of the end time.

generateAccessLogs_domainName :: Lens' GenerateAccessLogs Text

The name of the domain.

generateAccessLogs_appId :: Lens' GenerateAccessLogs Text

The unique ID for an Amplify app.

generateAccessLogsResponse_logUrl :: Lens' GenerateAccessLogsResponse (Maybe Text)

The pre-signed URL for the requested access logs.

generateAccessLogsResponse_httpStatus :: Lens' GenerateAccessLogsResponse Int

The response's http status code.

ListApps

listApps_nextToken :: Lens' ListApps (Maybe Text)

A pagination token. If non-null, the pagination token is returned in a result. Pass its value in another request to retrieve more entries.

listApps_maxResults :: Lens' ListApps (Maybe Natural)

The maximum number of records to list in a single response.

listAppsResponse_nextToken :: Lens' ListAppsResponse (Maybe Text)

A pagination token. Set to null to start listing apps from start. If non-null, the pagination token is returned in a result. Pass its value in here to list more projects.

listAppsResponse_httpStatus :: Lens' ListAppsResponse Int

The response's http status code.

listAppsResponse_apps :: Lens' ListAppsResponse [App]

A list of Amplify apps.

ListBranches

listBranches_nextToken :: Lens' ListBranches (Maybe Text)

A pagination token. Set to null to start listing branches from the start. If a non-null pagination token is returned in a result, pass its value in here to list more branches.

listBranches_maxResults :: Lens' ListBranches (Maybe Natural)

The maximum number of records to list in a single response.

listBranches_appId :: Lens' ListBranches Text

The unique ID for an Amplify app.

listBranchesResponse_nextToken :: Lens' ListBranchesResponse (Maybe Text)

A pagination token. If a non-null pagination token is returned in a result, pass its value in another request to retrieve more entries.

listBranchesResponse_httpStatus :: Lens' ListBranchesResponse Int

The response's http status code.

listBranchesResponse_branches :: Lens' ListBranchesResponse [Branch]

A list of branches for an Amplify app.

DeleteBackendEnvironment

deleteBackendEnvironment_appId :: Lens' DeleteBackendEnvironment Text

The unique ID of an Amplify app.

deleteBackendEnvironment_environmentName :: Lens' DeleteBackendEnvironment Text

The name of a backend environment of an Amplify app.

deleteBackendEnvironmentResponse_httpStatus :: Lens' DeleteBackendEnvironmentResponse Int

The response's http status code.

deleteBackendEnvironmentResponse_backendEnvironment :: Lens' DeleteBackendEnvironmentResponse BackendEnvironment

Describes the backend environment for an Amplify app.

DeleteApp

deleteApp_appId :: Lens' DeleteApp Text

The unique ID for an Amplify app.

deleteAppResponse_httpStatus :: Lens' DeleteAppResponse Int

The response's http status code.

deleteAppResponse_app :: Lens' DeleteAppResponse App

Undocumented member.

UpdateApp

updateApp_enableBranchAutoBuild :: Lens' UpdateApp (Maybe Bool)

Enables branch auto-building for an Amplify app.

updateApp_oauthToken :: Lens' UpdateApp (Maybe Text)

The OAuth token for a third-party source control system for an Amplify app. The token is used to create a webhook and a read-only deploy key. The OAuth token is not stored.

updateApp_accessToken :: Lens' UpdateApp (Maybe Text)

The personal access token for a third-party source control system for an Amplify app. The token is used to create webhook and a read-only deploy key. The token is not stored.

updateApp_customHeaders :: Lens' UpdateApp (Maybe Text)

The custom HTTP headers for an Amplify app.

updateApp_platform :: Lens' UpdateApp (Maybe Platform)

The platform for an Amplify app.

updateApp_basicAuthCredentials :: Lens' UpdateApp (Maybe Text)

The basic authorization credentials for an Amplify app.

updateApp_repository :: Lens' UpdateApp (Maybe Text)

The name of the repository for an Amplify app

updateApp_buildSpec :: Lens' UpdateApp (Maybe Text)

The build specification (build spec) for an Amplify app.

updateApp_enableBranchAutoDeletion :: Lens' UpdateApp (Maybe Bool)

Automatically disconnects a branch in the Amplify Console when you delete a branch from your Git repository.

updateApp_customRules :: Lens' UpdateApp (Maybe [CustomRule])

The custom redirect and rewrite rules for an Amplify app.

updateApp_iamServiceRoleArn :: Lens' UpdateApp (Maybe Text)

The AWS Identity and Access Management (IAM) service role for an Amplify app.

updateApp_autoBranchCreationPatterns :: Lens' UpdateApp (Maybe [Text])

Describes the automated branch creation glob patterns for an Amplify app.

updateApp_name :: Lens' UpdateApp (Maybe Text)

The name for an Amplify app.

updateApp_autoBranchCreationConfig :: Lens' UpdateApp (Maybe AutoBranchCreationConfig)

The automated branch creation configuration for an Amplify app.

updateApp_environmentVariables :: Lens' UpdateApp (Maybe (HashMap Text Text))

The environment variables for an Amplify app.

updateApp_enableAutoBranchCreation :: Lens' UpdateApp (Maybe Bool)

Enables automated branch creation for an Amplify app.

updateApp_enableBasicAuth :: Lens' UpdateApp (Maybe Bool)

Enables basic authorization for an Amplify app.

updateApp_description :: Lens' UpdateApp (Maybe Text)

The description for an Amplify app.

updateApp_appId :: Lens' UpdateApp Text

The unique ID for an Amplify app.

updateAppResponse_httpStatus :: Lens' UpdateAppResponse Int

The response's http status code.

updateAppResponse_app :: Lens' UpdateAppResponse App

Represents the updated Amplify app.

GetArtifactUrl

getArtifactUrl_artifactId :: Lens' GetArtifactUrl Text

The unique ID for an artifact.

getArtifactUrlResponse_httpStatus :: Lens' GetArtifactUrlResponse Int

The response's http status code.

getArtifactUrlResponse_artifactId :: Lens' GetArtifactUrlResponse Text

The unique ID for an artifact.

getArtifactUrlResponse_artifactUrl :: Lens' GetArtifactUrlResponse Text

The presigned URL for the artifact.

ListJobs

listJobs_nextToken :: Lens' ListJobs (Maybe Text)

A pagination token. Set to null to start listing steps from the start. If a non-null pagination token is returned in a result, pass its value in here to list more steps.

listJobs_maxResults :: Lens' ListJobs (Maybe Natural)

The maximum number of records to list in a single response.

listJobs_appId :: Lens' ListJobs Text

The unique ID for an Amplify app.

listJobs_branchName :: Lens' ListJobs Text

The name for a branch.

listJobsResponse_nextToken :: Lens' ListJobsResponse (Maybe Text)

A pagination token. If non-null the pagination token is returned in a result. Pass its value in another request to retrieve more entries.

listJobsResponse_httpStatus :: Lens' ListJobsResponse Int

The response's http status code.

listJobsResponse_jobSummaries :: Lens' ListJobsResponse [JobSummary]

The result structure for the list job result request.

DeleteJob

deleteJob_appId :: Lens' DeleteJob Text

The unique ID for an Amplify app.

deleteJob_branchName :: Lens' DeleteJob Text

The name for the branch, for the job.

deleteJob_jobId :: Lens' DeleteJob Text

The unique ID for the job.

deleteJobResponse_httpStatus :: Lens' DeleteJobResponse Int

The response's http status code.

deleteJobResponse_jobSummary :: Lens' DeleteJobResponse JobSummary

Undocumented member.

GetJob

getJob_appId :: Lens' GetJob Text

The unique ID for an Amplify app.

getJob_branchName :: Lens' GetJob Text

The branch name for the job.

getJob_jobId :: Lens' GetJob Text

The unique ID for the job.

getJobResponse_httpStatus :: Lens' GetJobResponse Int

The response's http status code.

getJobResponse_job :: Lens' GetJobResponse Job

Undocumented member.

StartJob

startJob_commitId :: Lens' StartJob (Maybe Text)

The commit ID from a third-party repository provider for the job.

startJob_jobId :: Lens' StartJob (Maybe Text)

The unique ID for an existing job. This is required if the value of jobType is RETRY.

startJob_jobReason :: Lens' StartJob (Maybe Text)

A descriptive reason for starting this job.

startJob_commitTime :: Lens' StartJob (Maybe UTCTime)

The commit date and time for the job.

startJob_commitMessage :: Lens' StartJob (Maybe Text)

The commit message from a third-party repository provider for the job.

startJob_appId :: Lens' StartJob Text

The unique ID for an Amplify app.

startJob_branchName :: Lens' StartJob Text

The branch name for the job.

startJob_jobType :: Lens' StartJob JobType

Describes the type for the job. The job type RELEASE starts a new job with the latest change from the specified branch. This value is available only for apps that are connected to a repository. The job type RETRY retries an existing job. If the job type value is RETRY, the jobId is also required.

startJobResponse_httpStatus :: Lens' StartJobResponse Int

The response's http status code.

startJobResponse_jobSummary :: Lens' StartJobResponse JobSummary

The summary for the job.

GetApp

getApp_appId :: Lens' GetApp Text

The unique ID for an Amplify app.

getAppResponse_httpStatus :: Lens' GetAppResponse Int

The response's http status code.

getAppResponse_app :: Lens' GetAppResponse App

Undocumented member.

UpdateWebhook

updateWebhook_branchName :: Lens' UpdateWebhook (Maybe Text)

The name for a branch that is part of an Amplify app.

updateWebhook_description :: Lens' UpdateWebhook (Maybe Text)

The description for a webhook.

updateWebhook_webhookId :: Lens' UpdateWebhook Text

The unique ID for a webhook.

updateWebhookResponse_httpStatus :: Lens' UpdateWebhookResponse Int

The response's http status code.

updateWebhookResponse_webhook :: Lens' UpdateWebhookResponse Webhook

Describes a webhook that connects repository events to an Amplify app.

DeleteWebhook

deleteWebhook_webhookId :: Lens' DeleteWebhook Text

The unique ID for a webhook.

deleteWebhookResponse_httpStatus :: Lens' DeleteWebhookResponse Int

The response's http status code.

deleteWebhookResponse_webhook :: Lens' DeleteWebhookResponse Webhook

Describes a webhook that connects repository events to an Amplify app.

ListWebhooks

listWebhooks_nextToken :: Lens' ListWebhooks (Maybe Text)

A pagination token. Set to null to start listing webhooks from the start. If non-null,the pagination token is returned in a result. Pass its value in here to list more webhooks.

listWebhooks_maxResults :: Lens' ListWebhooks (Maybe Natural)

The maximum number of records to list in a single response.

listWebhooks_appId :: Lens' ListWebhooks Text

The unique ID for an Amplify app.

listWebhooksResponse_nextToken :: Lens' ListWebhooksResponse (Maybe Text)

A pagination token. If non-null, the pagination token is returned in a result. Pass its value in another request to retrieve more entries.

listWebhooksResponse_httpStatus :: Lens' ListWebhooksResponse Int

The response's http status code.

listWebhooksResponse_webhooks :: Lens' ListWebhooksResponse [Webhook]

A list of webhooks.

CreateApp

createApp_enableBranchAutoBuild :: Lens' CreateApp (Maybe Bool)

Enables the auto building of branches for an Amplify app.

createApp_oauthToken :: Lens' CreateApp (Maybe Text)

The OAuth token for a third-party source control system for an Amplify app. The OAuth token is used to create a webhook and a read-only deploy key. The OAuth token is not stored.

createApp_accessToken :: Lens' CreateApp (Maybe Text)

The personal access token for a third-party source control system for an Amplify app. The personal access token is used to create a webhook and a read-only deploy key. The token is not stored.

createApp_customHeaders :: Lens' CreateApp (Maybe Text)

The custom HTTP headers for an Amplify app.

createApp_platform :: Lens' CreateApp (Maybe Platform)

The platform or framework for an Amplify app.

createApp_basicAuthCredentials :: Lens' CreateApp (Maybe Text)

The credentials for basic authorization for an Amplify app.

createApp_repository :: Lens' CreateApp (Maybe Text)

The repository for an Amplify app.

createApp_buildSpec :: Lens' CreateApp (Maybe Text)

The build specification (build spec) for an Amplify app.

createApp_enableBranchAutoDeletion :: Lens' CreateApp (Maybe Bool)

Automatically disconnects a branch in the Amplify Console when you delete a branch from your Git repository.

createApp_customRules :: Lens' CreateApp (Maybe [CustomRule])

The custom rewrite and redirect rules for an Amplify app.

createApp_iamServiceRoleArn :: Lens' CreateApp (Maybe Text)

The AWS Identity and Access Management (IAM) service role for an Amplify app.

createApp_autoBranchCreationPatterns :: Lens' CreateApp (Maybe [Text])

The automated branch creation glob patterns for an Amplify app.

createApp_autoBranchCreationConfig :: Lens' CreateApp (Maybe AutoBranchCreationConfig)

The automated branch creation configuration for an Amplify app.

createApp_environmentVariables :: Lens' CreateApp (Maybe (HashMap Text Text))

The environment variables map for an Amplify app.

createApp_enableAutoBranchCreation :: Lens' CreateApp (Maybe Bool)

Enables automated branch creation for an Amplify app.

createApp_enableBasicAuth :: Lens' CreateApp (Maybe Bool)

Enables basic authorization for an Amplify app. This will apply to all branches that are part of this app.

createApp_description :: Lens' CreateApp (Maybe Text)

The description for an Amplify app.

createApp_tags :: Lens' CreateApp (Maybe (HashMap Text Text))

The tag for an Amplify app.

createApp_name :: Lens' CreateApp Text

The name for an Amplify app.

createAppResponse_httpStatus :: Lens' CreateAppResponse Int

The response's http status code.

createAppResponse_app :: Lens' CreateAppResponse App

Undocumented member.

DeleteDomainAssociation

deleteDomainAssociation_appId :: Lens' DeleteDomainAssociation Text

The unique id for an Amplify app.

deleteDomainAssociation_domainName :: Lens' DeleteDomainAssociation Text

The name of the domain.

deleteDomainAssociationResponse_httpStatus :: Lens' DeleteDomainAssociationResponse Int

The response's http status code.

deleteDomainAssociationResponse_domainAssociation :: Lens' DeleteDomainAssociationResponse DomainAssociation

Undocumented member.

UpdateDomainAssociation

updateDomainAssociation_enableAutoSubDomain :: Lens' UpdateDomainAssociation (Maybe Bool)

Enables the automated creation of subdomains for branches.

updateDomainAssociation_autoSubDomainCreationPatterns :: Lens' UpdateDomainAssociation (Maybe [Text])

Sets the branch patterns for automatic subdomain creation.

updateDomainAssociation_autoSubDomainIAMRole :: Lens' UpdateDomainAssociation (Maybe Text)

The required AWS Identity and Access Management (IAM) service role for the Amazon Resource Name (ARN) for automatically creating subdomains.

updateDomainAssociation_appId :: Lens' UpdateDomainAssociation Text

The unique ID for an Amplify app.

updateDomainAssociation_domainName :: Lens' UpdateDomainAssociation Text

The name of the domain.

updateDomainAssociation_subDomainSettings :: Lens' UpdateDomainAssociation [SubDomainSetting]

Describes the settings for the subdomain.

updateDomainAssociationResponse_httpStatus :: Lens' UpdateDomainAssociationResponse Int

The response's http status code.

updateDomainAssociationResponse_domainAssociation :: Lens' UpdateDomainAssociationResponse DomainAssociation

Describes a domain association, which associates a custom domain with an Amplify app.

ListDomainAssociations

listDomainAssociations_nextToken :: Lens' ListDomainAssociations (Maybe Text)

A pagination token. Set to null to start listing apps from the start. If non-null, a pagination token is returned in a result. Pass its value in here to list more projects.

listDomainAssociations_maxResults :: Lens' ListDomainAssociations (Maybe Natural)

The maximum number of records to list in a single response.

listDomainAssociations_appId :: Lens' ListDomainAssociations Text

The unique ID for an Amplify app.

listDomainAssociationsResponse_nextToken :: Lens' ListDomainAssociationsResponse (Maybe Text)

A pagination token. If non-null, a pagination token is returned in a result. Pass its value in another request to retrieve more entries.

listDomainAssociationsResponse_httpStatus :: Lens' ListDomainAssociationsResponse Int

The response's http status code.

listDomainAssociationsResponse_domainAssociations :: Lens' ListDomainAssociationsResponse [DomainAssociation]

A list of domain associations.

TagResource

tagResource_resourceArn :: Lens' TagResource Text

The Amazon Resource Name (ARN) to use to tag a resource.

tagResource_tags :: Lens' TagResource (HashMap Text Text)

The tags used to tag the resource.

tagResourceResponse_httpStatus :: Lens' TagResourceResponse Int

The response's http status code.

ListBackendEnvironments

listBackendEnvironments_nextToken :: Lens' ListBackendEnvironments (Maybe Text)

A pagination token. Set to null to start listing backend environments from the start. If a non-null pagination token is returned in a result, pass its value in here to list more backend environments.

listBackendEnvironments_environmentName :: Lens' ListBackendEnvironments (Maybe Text)

The name of the backend environment

listBackendEnvironments_maxResults :: Lens' ListBackendEnvironments (Maybe Natural)

The maximum number of records to list in a single response.

listBackendEnvironments_appId :: Lens' ListBackendEnvironments Text

The unique ID for an Amplify app.

listBackendEnvironmentsResponse_nextToken :: Lens' ListBackendEnvironmentsResponse (Maybe Text)

A pagination token. If a non-null pagination token is returned in a result, pass its value in another request to retrieve more entries.

listBackendEnvironmentsResponse_httpStatus :: Lens' ListBackendEnvironmentsResponse Int

The response's http status code.

listBackendEnvironmentsResponse_backendEnvironments :: Lens' ListBackendEnvironmentsResponse [BackendEnvironment]

The list of backend environments for an Amplify app.

UntagResource

untagResource_resourceArn :: Lens' UntagResource Text

The Amazon Resource Name (ARN) to use to untag a resource.

untagResource_tagKeys :: Lens' UntagResource (NonEmpty Text)

The tag keys to use to untag a resource.

untagResourceResponse_httpStatus :: Lens' UntagResourceResponse Int

The response's http status code.

StartDeployment

startDeployment_jobId :: Lens' StartDeployment (Maybe Text)

The job ID for this deployment, generated by the create deployment request.

startDeployment_sourceUrl :: Lens' StartDeployment (Maybe Text)

The source URL for this deployment, used when calling start deployment without create deployment. The source URL can be any HTTP GET URL that is publicly accessible and downloads a single .zip file.

startDeployment_appId :: Lens' StartDeployment Text

The unique ID for an Amplify app.

startDeployment_branchName :: Lens' StartDeployment Text

The name for the branch, for the job.

startDeploymentResponse_httpStatus :: Lens' StartDeploymentResponse Int

The response's http status code.

startDeploymentResponse_jobSummary :: Lens' StartDeploymentResponse JobSummary

The summary for the job.

Types

App

app_customHeaders :: Lens' App (Maybe Text)

Describes the custom HTTP headers for the Amplify app.

app_basicAuthCredentials :: Lens' App (Maybe Text)

The basic authorization credentials for branches for the Amplify app.

app_buildSpec :: Lens' App (Maybe Text)

Describes the content of the build specification (build spec) for the Amplify app.

app_enableBranchAutoDeletion :: Lens' App (Maybe Bool)

Automatically disconnect a branch in the Amplify Console when you delete a branch from your Git repository.

app_customRules :: Lens' App (Maybe [CustomRule])

Describes the custom redirect and rewrite rules for the Amplify app.

app_iamServiceRoleArn :: Lens' App (Maybe Text)

The AWS Identity and Access Management (IAM) service role for the Amazon Resource Name (ARN) of the Amplify app.

app_autoBranchCreationPatterns :: Lens' App (Maybe [Text])

Describes the automated branch creation glob patterns for the Amplify app.

app_productionBranch :: Lens' App (Maybe ProductionBranch)

Describes the information about a production branch of the Amplify app.

app_autoBranchCreationConfig :: Lens' App (Maybe AutoBranchCreationConfig)

Describes the automated branch creation configuration for the Amplify app.

app_enableAutoBranchCreation :: Lens' App (Maybe Bool)

Enables automated branch creation for the Amplify app.

app_tags :: Lens' App (Maybe (HashMap Text Text))

The tag for the Amplify app.

app_appId :: Lens' App Text

The unique ID of the Amplify app.

app_appArn :: Lens' App Text

The Amazon Resource Name (ARN) of the Amplify app.

app_name :: Lens' App Text

The name for the Amplify app.

app_description :: Lens' App Text

The description for the Amplify app.

app_repository :: Lens' App Text

The repository for the Amplify app.

app_platform :: Lens' App Platform

The platform for the Amplify app.

app_createTime :: Lens' App UTCTime

Creates a date and time for the Amplify app.

app_updateTime :: Lens' App UTCTime

Updates the date and time for the Amplify app.

app_environmentVariables :: Lens' App (HashMap Text Text)

The environment variables for the Amplify app.

app_defaultDomain :: Lens' App Text

The default domain for the Amplify app.

app_enableBranchAutoBuild :: Lens' App Bool

Enables the auto-building of branches for the Amplify app.

app_enableBasicAuth :: Lens' App Bool

Enables basic authorization for the Amplify app's branches.

Artifact

artifact_artifactFileName :: Lens' Artifact Text

The file name for the artifact.

artifact_artifactId :: Lens' Artifact Text

The unique ID for the artifact.

AutoBranchCreationConfig

autoBranchCreationConfig_framework :: Lens' AutoBranchCreationConfig (Maybe Text)

The framework for the autocreated branch.

autoBranchCreationConfig_stage :: Lens' AutoBranchCreationConfig (Maybe Stage)

Describes the current stage for the autocreated branch.

autoBranchCreationConfig_enablePullRequestPreview :: Lens' AutoBranchCreationConfig (Maybe Bool)

Enables pull request previews for the autocreated branch.

autoBranchCreationConfig_basicAuthCredentials :: Lens' AutoBranchCreationConfig (Maybe Text)

The basic authorization credentials for the autocreated branch.

autoBranchCreationConfig_buildSpec :: Lens' AutoBranchCreationConfig (Maybe Text)

The build specification (build spec) for the autocreated branch.

autoBranchCreationConfig_enablePerformanceMode :: Lens' AutoBranchCreationConfig (Maybe Bool)

Enables performance mode for the branch.

Performance mode optimizes for faster hosting performance by keeping content cached at the edge for a longer interval. When performance mode is enabled, hosting configuration or code changes can take up to 10 minutes to roll out.

autoBranchCreationConfig_environmentVariables :: Lens' AutoBranchCreationConfig (Maybe (HashMap Text Text))

The environment variables for the autocreated branch.

autoBranchCreationConfig_enableAutoBuild :: Lens' AutoBranchCreationConfig (Maybe Bool)

Enables auto building for the autocreated branch.

autoBranchCreationConfig_enableBasicAuth :: Lens' AutoBranchCreationConfig (Maybe Bool)

Enables basic authorization for the autocreated branch.

autoBranchCreationConfig_pullRequestEnvironmentName :: Lens' AutoBranchCreationConfig (Maybe Text)

The Amplify environment name for the pull request.

BackendEnvironment

backendEnvironment_deploymentArtifacts :: Lens' BackendEnvironment (Maybe Text)

The name of deployment artifacts.

backendEnvironment_stackName :: Lens' BackendEnvironment (Maybe Text)

The AWS CloudFormation stack name of a backend environment.

backendEnvironment_backendEnvironmentArn :: Lens' BackendEnvironment Text

The Amazon Resource Name (ARN) for a backend environment that is part of an Amplify app.

backendEnvironment_environmentName :: Lens' BackendEnvironment Text

The name for a backend environment that is part of an Amplify app.

backendEnvironment_createTime :: Lens' BackendEnvironment UTCTime

The creation date and time for a backend environment that is part of an Amplify app.

backendEnvironment_updateTime :: Lens' BackendEnvironment UTCTime

The last updated date and time for a backend environment that is part of an Amplify app.

Branch

branch_backendEnvironmentArn :: Lens' Branch (Maybe Text)

The Amazon Resource Name (ARN) for a backend environment that is part of an Amplify app.

branch_thumbnailUrl :: Lens' Branch (Maybe Text)

The thumbnail URL for the branch of an Amplify app.

branch_basicAuthCredentials :: Lens' Branch (Maybe Text)

The basic authorization credentials for a branch of an Amplify app.

branch_buildSpec :: Lens' Branch (Maybe Text)

The build specification (build spec) content for the branch of an Amplify app.

branch_sourceBranch :: Lens' Branch (Maybe Text)

The source branch if the branch is a pull request branch.

branch_enablePerformanceMode :: Lens' Branch (Maybe Bool)

Enables performance mode for the branch.

Performance mode optimizes for faster hosting performance by keeping content cached at the edge for a longer interval. When performance mode is enabled, hosting configuration or code changes can take up to 10 minutes to roll out.

branch_destinationBranch :: Lens' Branch (Maybe Text)

The destination branch if the branch is a pull request branch.

branch_pullRequestEnvironmentName :: Lens' Branch (Maybe Text)

The Amplify environment name for the pull request.

branch_associatedResources :: Lens' Branch (Maybe [Text])

A list of custom resources that are linked to this branch.

branch_tags :: Lens' Branch (Maybe (HashMap Text Text))

The tag for the branch of an Amplify app.

branch_branchArn :: Lens' Branch Text

The Amazon Resource Name (ARN) for a branch that is part of an Amplify app.

branch_branchName :: Lens' Branch Text

The name for the branch that is part of an Amplify app.

branch_description :: Lens' Branch Text

The description for the branch that is part of an Amplify app.

branch_stage :: Lens' Branch Stage

The current stage for the branch that is part of an Amplify app.

branch_displayName :: Lens' Branch Text

The display name for the branch. This is used as the default domain prefix.

branch_enableNotification :: Lens' Branch Bool

Enables notifications for a branch that is part of an Amplify app.

branch_createTime :: Lens' Branch UTCTime

The creation date and time for a branch that is part of an Amplify app.

branch_updateTime :: Lens' Branch UTCTime

The last updated date and time for a branch that is part of an Amplify app.

branch_environmentVariables :: Lens' Branch (HashMap Text Text)

The environment variables specific to a branch of an Amplify app.

branch_enableAutoBuild :: Lens' Branch Bool

Enables auto-building on push for a branch of an Amplify app.

branch_customDomains :: Lens' Branch [Text]

The custom domains for a branch of an Amplify app.

branch_framework :: Lens' Branch Text

The framework for a branch of an Amplify app.

branch_activeJobId :: Lens' Branch Text

The ID of the active job for a branch of an Amplify app.

branch_totalNumberOfJobs :: Lens' Branch Text

The total number of jobs that are part of an Amplify app.

branch_enableBasicAuth :: Lens' Branch Bool

Enables basic authorization for a branch of an Amplify app.

branch_ttl :: Lens' Branch Text

The content Time to Live (TTL) for the website in seconds.

branch_enablePullRequestPreview :: Lens' Branch Bool

Enables pull request previews for the branch.

CustomRule

customRule_status :: Lens' CustomRule (Maybe Text)

The status code for a URL rewrite or redirect rule.

200

Represents a 200 rewrite rule.

301

Represents a 301 (moved pemanently) redirect rule. This and all future requests should be directed to the target URL.

302

Represents a 302 temporary redirect rule.

404

Represents a 404 redirect rule.

404-200

Represents a 404 rewrite rule.

customRule_condition :: Lens' CustomRule (Maybe Text)

The condition for a URL rewrite or redirect rule, such as a country code.

customRule_source :: Lens' CustomRule Text

The source pattern for a URL rewrite or redirect rule.

customRule_target :: Lens' CustomRule Text

The target pattern for a URL rewrite or redirect rule.

DomainAssociation

domainAssociation_certificateVerificationDNSRecord :: Lens' DomainAssociation (Maybe Text)

The DNS record for certificate verification.

domainAssociation_autoSubDomainCreationPatterns :: Lens' DomainAssociation (Maybe [Text])

Sets branch patterns for automatic subdomain creation.

domainAssociation_autoSubDomainIAMRole :: Lens' DomainAssociation (Maybe Text)

The required AWS Identity and Access Management (IAM) service role for the Amazon Resource Name (ARN) for automatically creating subdomains.

domainAssociation_domainAssociationArn :: Lens' DomainAssociation Text

The Amazon Resource Name (ARN) for the domain association.

domainAssociation_domainName :: Lens' DomainAssociation Text

The name of the domain.

domainAssociation_enableAutoSubDomain :: Lens' DomainAssociation Bool

Enables the automated creation of subdomains for branches.

domainAssociation_domainStatus :: Lens' DomainAssociation DomainStatus

The current status of the domain association.

domainAssociation_statusReason :: Lens' DomainAssociation Text

The reason for the current status of the domain association.

domainAssociation_subDomains :: Lens' DomainAssociation [SubDomain]

The subdomains for the domain association.

Job

job_summary :: Lens' Job JobSummary

Describes the summary for an execution job for an Amplify app.

job_steps :: Lens' Job [Step]

The execution steps for an execution job, for an Amplify app.

JobSummary

jobSummary_endTime :: Lens' JobSummary (Maybe UTCTime)

The end date and time for the job.

jobSummary_jobArn :: Lens' JobSummary Text

The Amazon Resource Name (ARN) for the job.

jobSummary_jobId :: Lens' JobSummary Text

The unique ID for the job.

jobSummary_commitId :: Lens' JobSummary Text

The commit ID from a third-party repository provider for the job.

jobSummary_commitMessage :: Lens' JobSummary Text

The commit message from a third-party repository provider for the job.

jobSummary_commitTime :: Lens' JobSummary UTCTime

The commit date and time for the job.

jobSummary_startTime :: Lens' JobSummary UTCTime

The start date and time for the job.

jobSummary_status :: Lens' JobSummary JobStatus

The current status for the job.

jobSummary_jobType :: Lens' JobSummary JobType

The type for the job. If the value is RELEASE, the job was manually released from its source by using the StartJob API. If the value is RETRY, the job was manually retried using the StartJob API. If the value is WEB_HOOK, the job was automatically triggered by webhooks.

ProductionBranch

productionBranch_lastDeployTime :: Lens' ProductionBranch (Maybe UTCTime)

The last deploy time of the production branch.

productionBranch_status :: Lens' ProductionBranch (Maybe Text)

The status of the production branch.

productionBranch_thumbnailUrl :: Lens' ProductionBranch (Maybe Text)

The thumbnail URL for the production branch.

productionBranch_branchName :: Lens' ProductionBranch (Maybe Text)

The branch name for the production branch.

Step

step_logUrl :: Lens' Step (Maybe Text)

The URL to the logs for the execution step.

step_context :: Lens' Step (Maybe Text)

The context for the current step. Includes a build image if the step is build.

step_testArtifactsUrl :: Lens' Step (Maybe Text)

The URL to the test artifact for the execution step.

step_artifactsUrl :: Lens' Step (Maybe Text)

The URL to the artifact for the execution step.

step_testConfigUrl :: Lens' Step (Maybe Text)

The URL to the test configuration for the execution step.

step_screenshots :: Lens' Step (Maybe (HashMap Text Text))

The list of screenshot URLs for the execution step, if relevant.

step_statusReason :: Lens' Step (Maybe Text)

The reason for the current step status.

step_stepName :: Lens' Step Text

The name of the execution step.

step_startTime :: Lens' Step UTCTime

The start date and time of the execution step.

step_status :: Lens' Step JobStatus

The status of the execution step.

step_endTime :: Lens' Step UTCTime

The end date and time of the execution step.

SubDomain

subDomain_subDomainSetting :: Lens' SubDomain SubDomainSetting

Describes the settings for the subdomain.

subDomain_verified :: Lens' SubDomain Bool

The verified status of the subdomain

subDomain_dnsRecord :: Lens' SubDomain Text

The DNS record for the subdomain.

SubDomainSetting

subDomainSetting_prefix :: Lens' SubDomainSetting Text

The prefix setting for the subdomain.

subDomainSetting_branchName :: Lens' SubDomainSetting Text

The branch name setting for the subdomain.

Webhook

webhook_webhookArn :: Lens' Webhook Text

The Amazon Resource Name (ARN) for the webhook.

webhook_webhookId :: Lens' Webhook Text

The ID of the webhook.

webhook_webhookUrl :: Lens' Webhook Text

The URL of the webhook.

webhook_branchName :: Lens' Webhook Text

The name for a branch that is part of an Amplify app.

webhook_description :: Lens' Webhook Text

The description for a webhook.

webhook_createTime :: Lens' Webhook UTCTime

The create date and time for a webhook.

webhook_updateTime :: Lens' Webhook UTCTime

Updates the date and time for a webhook.