Amazonka.Amplify.StartJob

Description

Starts a new job for a branch of an Amplify app.

Creating a Request

data StartJob

The request structure for the start job request.

See: newStartJob smart constructor.

Constructors

StartJob'

Fields

commitId :: Maybe Text

The commit ID from a third-party repository provider for the job.

jobId :: Maybe Text

The unique ID for an existing job. This is required if the value of jobType is RETRY.

jobReason :: Maybe Text

A descriptive reason for starting this job.

commitTime :: Maybe POSIX

The commit date and time for the job.

commitMessage :: Maybe Text

The commit message from a third-party repository provider for the job.

appId :: Text

The unique ID for an Amplify app.

branchName :: Text

The branch name for the job.

jobType :: JobType

Describes the type for the job. The job type RELEASE starts a new job with the latest change from the specified branch. This value is available only for apps that are connected to a repository. The job type RETRY retries an existing job. If the job type value is RETRY, the jobId is also required.

newStartJob

Arguments

:: Text
$sel:appId:StartJob'

-> Text
$sel:branchName:StartJob'

-> JobType
$sel:jobType:StartJob'

-> StartJob

Create a value of StartJob with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:commitId:StartJob', startJob_commitId - The commit ID from a third-party repository provider for the job.

$sel:jobId:StartJob', startJob_jobId - The unique ID for an existing job. This is required if the value of jobType is RETRY.

$sel:jobReason:StartJob', startJob_jobReason - A descriptive reason for starting this job.

$sel:commitTime:StartJob', startJob_commitTime - The commit date and time for the job.

$sel:commitMessage:StartJob', startJob_commitMessage - The commit message from a third-party repository provider for the job.

$sel:appId:StartJob', startJob_appId - The unique ID for an Amplify app.

$sel:branchName:StartJob', startJob_branchName - The branch name for the job.

$sel:jobType:StartJob', startJob_jobType - Describes the type for the job. The job type RELEASE starts a new job with the latest change from the specified branch. This value is available only for apps that are connected to a repository. The job type RETRY retries an existing job. If the job type value is RETRY, the jobId is also required.

Request Lenses

startJob_commitId :: Lens' StartJob (Maybe Text)

The commit ID from a third-party repository provider for the job.

startJob_jobId :: Lens' StartJob (Maybe Text)

The unique ID for an existing job. This is required if the value of jobType is RETRY.

startJob_jobReason :: Lens' StartJob (Maybe Text)

A descriptive reason for starting this job.

startJob_commitTime :: Lens' StartJob (Maybe UTCTime)

The commit date and time for the job.

startJob_commitMessage :: Lens' StartJob (Maybe Text)

The commit message from a third-party repository provider for the job.

startJob_appId :: Lens' StartJob Text

The unique ID for an Amplify app.

startJob_branchName :: Lens' StartJob Text

The branch name for the job.

startJob_jobType :: Lens' StartJob JobType

Describes the type for the job. The job type RELEASE starts a new job with the latest change from the specified branch. This value is available only for apps that are connected to a repository. The job type RETRY retries an existing job. If the job type value is RETRY, the jobId is also required.

Destructuring the Response

data StartJobResponse

The result structure for the run job request.

See: newStartJobResponse smart constructor.

Constructors

StartJobResponse'

Fields

httpStatus :: Int

The response's http status code.

jobSummary :: JobSummary

The summary for the job.

newStartJobResponse

Arguments

:: Int
$sel:httpStatus:StartJobResponse'

-> JobSummary
$sel:jobSummary:StartJobResponse'

-> StartJobResponse

Create a value of StartJobResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:httpStatus:StartJobResponse', startJobResponse_httpStatus - The response's http status code.

$sel:jobSummary:StartJobResponse', startJobResponse_jobSummary - The summary for the job.

Response Lenses

startJobResponse_httpStatus :: Lens' StartJobResponse Int

The response's http status code.

startJobResponse_jobSummary :: Lens' StartJobResponse JobSummary

The summary for the job.