Amazonka.Amplify.Types.App

Description

Documentation

data App

Represents the different branches of a repository for building, deploying, and hosting an Amplify app.

See: newApp smart constructor.

Constructors

App'

Fields

customHeaders :: Maybe Text

Describes the custom HTTP headers for the Amplify app.

basicAuthCredentials :: Maybe (Sensitive Text)

The basic authorization credentials for branches for the Amplify app.

buildSpec :: Maybe Text

Describes the content of the build specification (build spec) for the Amplify app.

enableBranchAutoDeletion :: Maybe Bool

Automatically disconnect a branch in the Amplify Console when you delete a branch from your Git repository.

customRules :: Maybe [CustomRule]

Describes the custom redirect and rewrite rules for the Amplify app.

iamServiceRoleArn :: Maybe Text

The AWS Identity and Access Management (IAM) service role for the Amazon Resource Name (ARN) of the Amplify app.

autoBranchCreationPatterns :: Maybe [Text]

Describes the automated branch creation glob patterns for the Amplify app.

productionBranch :: Maybe ProductionBranch

Describes the information about a production branch of the Amplify app.

autoBranchCreationConfig :: Maybe AutoBranchCreationConfig

Describes the automated branch creation configuration for the Amplify app.

enableAutoBranchCreation :: Maybe Bool

Enables automated branch creation for the Amplify app.

tags :: Maybe (HashMap Text Text)

The tag for the Amplify app.

appId :: Text

The unique ID of the Amplify app.

appArn :: Text

The Amazon Resource Name (ARN) of the Amplify app.

name :: Text

The name for the Amplify app.

description :: Text

The description for the Amplify app.

repository :: Text

The repository for the Amplify app.

platform :: Platform

The platform for the Amplify app.

createTime :: POSIX

Creates a date and time for the Amplify app.

updateTime :: POSIX

Updates the date and time for the Amplify app.

environmentVariables :: HashMap Text Text

The environment variables for the Amplify app.

defaultDomain :: Text

The default domain for the Amplify app.

enableBranchAutoBuild :: Bool

Enables the auto-building of branches for the Amplify app.

enableBasicAuth :: Bool

Enables basic authorization for the Amplify app's branches.

newApp

Arguments

:: Text
$sel:appId:App'

-> Text
$sel:appArn:App'

-> Text
$sel:name:App'

-> Text
$sel:description:App'

-> Text
$sel:repository:App'

-> Platform
$sel:platform:App'

-> UTCTime
$sel:createTime:App'

-> UTCTime
$sel:updateTime:App'

-> Text
$sel:defaultDomain:App'

-> Bool
$sel:enableBranchAutoBuild:App'

-> Bool
$sel:enableBasicAuth:App'

-> App

Create a value of App with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:customHeaders:App', app_customHeaders - Describes the custom HTTP headers for the Amplify app.

$sel:basicAuthCredentials:App', app_basicAuthCredentials - The basic authorization credentials for branches for the Amplify app.

$sel:buildSpec:App', app_buildSpec - Describes the content of the build specification (build spec) for the Amplify app.

$sel:enableBranchAutoDeletion:App', app_enableBranchAutoDeletion - Automatically disconnect a branch in the Amplify Console when you delete a branch from your Git repository.

$sel:customRules:App', app_customRules - Describes the custom redirect and rewrite rules for the Amplify app.

$sel:iamServiceRoleArn:App', app_iamServiceRoleArn - The AWS Identity and Access Management (IAM) service role for the Amazon Resource Name (ARN) of the Amplify app.

$sel:autoBranchCreationPatterns:App', app_autoBranchCreationPatterns - Describes the automated branch creation glob patterns for the Amplify app.

$sel:productionBranch:App', app_productionBranch - Describes the information about a production branch of the Amplify app.

$sel:autoBranchCreationConfig:App', app_autoBranchCreationConfig - Describes the automated branch creation configuration for the Amplify app.

$sel:enableAutoBranchCreation:App', app_enableAutoBranchCreation - Enables automated branch creation for the Amplify app.

$sel:tags:App', app_tags - The tag for the Amplify app.

$sel:appId:App', app_appId - The unique ID of the Amplify app.

$sel:appArn:App', app_appArn - The Amazon Resource Name (ARN) of the Amplify app.

$sel:name:App', app_name - The name for the Amplify app.

$sel:description:App', app_description - The description for the Amplify app.

$sel:repository:App', app_repository - The repository for the Amplify app.

$sel:platform:App', app_platform - The platform for the Amplify app.

$sel:createTime:App', app_createTime - Creates a date and time for the Amplify app.

$sel:updateTime:App', app_updateTime - Updates the date and time for the Amplify app.

$sel:environmentVariables:App', app_environmentVariables - The environment variables for the Amplify app.

$sel:defaultDomain:App', app_defaultDomain - The default domain for the Amplify app.

$sel:enableBranchAutoBuild:App', app_enableBranchAutoBuild - Enables the auto-building of branches for the Amplify app.

$sel:enableBasicAuth:App', app_enableBasicAuth - Enables basic authorization for the Amplify app's branches.

app_customHeaders :: Lens' App (Maybe Text)

Describes the custom HTTP headers for the Amplify app.

app_basicAuthCredentials :: Lens' App (Maybe Text)

The basic authorization credentials for branches for the Amplify app.

app_buildSpec :: Lens' App (Maybe Text)

Describes the content of the build specification (build spec) for the Amplify app.

app_enableBranchAutoDeletion :: Lens' App (Maybe Bool)

Automatically disconnect a branch in the Amplify Console when you delete a branch from your Git repository.

app_customRules :: Lens' App (Maybe [CustomRule])

Describes the custom redirect and rewrite rules for the Amplify app.

app_iamServiceRoleArn :: Lens' App (Maybe Text)

The AWS Identity and Access Management (IAM) service role for the Amazon Resource Name (ARN) of the Amplify app.

app_autoBranchCreationPatterns :: Lens' App (Maybe [Text])

Describes the automated branch creation glob patterns for the Amplify app.

app_productionBranch :: Lens' App (Maybe ProductionBranch)

Describes the information about a production branch of the Amplify app.

app_autoBranchCreationConfig :: Lens' App (Maybe AutoBranchCreationConfig)

Describes the automated branch creation configuration for the Amplify app.

app_enableAutoBranchCreation :: Lens' App (Maybe Bool)

Enables automated branch creation for the Amplify app.

app_tags :: Lens' App (Maybe (HashMap Text Text))

The tag for the Amplify app.

app_appId :: Lens' App Text

The unique ID of the Amplify app.

app_appArn :: Lens' App Text

The Amazon Resource Name (ARN) of the Amplify app.

app_name :: Lens' App Text

The name for the Amplify app.

app_description :: Lens' App Text

The description for the Amplify app.

app_repository :: Lens' App Text

The repository for the Amplify app.

app_platform :: Lens' App Platform

The platform for the Amplify app.

app_createTime :: Lens' App UTCTime

Creates a date and time for the Amplify app.

app_updateTime :: Lens' App UTCTime

Updates the date and time for the Amplify app.

app_environmentVariables :: Lens' App (HashMap Text Text)

The environment variables for the Amplify app.

app_defaultDomain :: Lens' App Text

The default domain for the Amplify app.

app_enableBranchAutoBuild :: Lens' App Bool

Enables the auto-building of branches for the Amplify app.

app_enableBasicAuth :: Lens' App Bool

Enables basic authorization for the Amplify app's branches.