Amazonka.Amplify.Types.AutoBranchCreationConfig

Description

Documentation

data AutoBranchCreationConfig

Describes the automated branch creation configuration.

See: newAutoBranchCreationConfig smart constructor.

Constructors

AutoBranchCreationConfig'

Fields

framework :: Maybe Text

The framework for the autocreated branch.

stage :: Maybe Stage

Describes the current stage for the autocreated branch.

enablePullRequestPreview :: Maybe Bool

Enables pull request previews for the autocreated branch.

basicAuthCredentials :: Maybe (Sensitive Text)

The basic authorization credentials for the autocreated branch.

buildSpec :: Maybe Text

The build specification (build spec) for the autocreated branch.

enablePerformanceMode :: Maybe Bool

Enables performance mode for the branch.

Performance mode optimizes for faster hosting performance by keeping content cached at the edge for a longer interval. When performance mode is enabled, hosting configuration or code changes can take up to 10 minutes to roll out.

environmentVariables :: Maybe (HashMap Text Text)

The environment variables for the autocreated branch.

enableAutoBuild :: Maybe Bool

Enables auto building for the autocreated branch.

enableBasicAuth :: Maybe Bool

Enables basic authorization for the autocreated branch.

pullRequestEnvironmentName :: Maybe Text

The Amplify environment name for the pull request.

newAutoBranchCreationConfig :: AutoBranchCreationConfig

Create a value of AutoBranchCreationConfig with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:framework:AutoBranchCreationConfig', autoBranchCreationConfig_framework - The framework for the autocreated branch.

$sel:stage:AutoBranchCreationConfig', autoBranchCreationConfig_stage - Describes the current stage for the autocreated branch.

$sel:enablePullRequestPreview:AutoBranchCreationConfig', autoBranchCreationConfig_enablePullRequestPreview - Enables pull request previews for the autocreated branch.

$sel:basicAuthCredentials:AutoBranchCreationConfig', autoBranchCreationConfig_basicAuthCredentials - The basic authorization credentials for the autocreated branch.

$sel:buildSpec:AutoBranchCreationConfig', autoBranchCreationConfig_buildSpec - The build specification (build spec) for the autocreated branch.

$sel:enablePerformanceMode:AutoBranchCreationConfig', autoBranchCreationConfig_enablePerformanceMode - Enables performance mode for the branch.

Performance mode optimizes for faster hosting performance by keeping content cached at the edge for a longer interval. When performance mode is enabled, hosting configuration or code changes can take up to 10 minutes to roll out.

$sel:environmentVariables:AutoBranchCreationConfig', autoBranchCreationConfig_environmentVariables - The environment variables for the autocreated branch.

$sel:enableAutoBuild:AutoBranchCreationConfig', autoBranchCreationConfig_enableAutoBuild - Enables auto building for the autocreated branch.

$sel:enableBasicAuth:AutoBranchCreationConfig', autoBranchCreationConfig_enableBasicAuth - Enables basic authorization for the autocreated branch.

$sel:pullRequestEnvironmentName:AutoBranchCreationConfig', autoBranchCreationConfig_pullRequestEnvironmentName - The Amplify environment name for the pull request.

autoBranchCreationConfig_framework :: Lens' AutoBranchCreationConfig (Maybe Text)

The framework for the autocreated branch.

autoBranchCreationConfig_stage :: Lens' AutoBranchCreationConfig (Maybe Stage)

Describes the current stage for the autocreated branch.

autoBranchCreationConfig_enablePullRequestPreview :: Lens' AutoBranchCreationConfig (Maybe Bool)

Enables pull request previews for the autocreated branch.

autoBranchCreationConfig_basicAuthCredentials :: Lens' AutoBranchCreationConfig (Maybe Text)

The basic authorization credentials for the autocreated branch.

autoBranchCreationConfig_buildSpec :: Lens' AutoBranchCreationConfig (Maybe Text)

The build specification (build spec) for the autocreated branch.

autoBranchCreationConfig_enablePerformanceMode :: Lens' AutoBranchCreationConfig (Maybe Bool)

Enables performance mode for the branch.

Performance mode optimizes for faster hosting performance by keeping content cached at the edge for a longer interval. When performance mode is enabled, hosting configuration or code changes can take up to 10 minutes to roll out.

autoBranchCreationConfig_environmentVariables :: Lens' AutoBranchCreationConfig (Maybe (HashMap Text Text))

The environment variables for the autocreated branch.

autoBranchCreationConfig_enableAutoBuild :: Lens' AutoBranchCreationConfig (Maybe Bool)

Enables auto building for the autocreated branch.

autoBranchCreationConfig_enableBasicAuth :: Lens' AutoBranchCreationConfig (Maybe Bool)

Enables basic authorization for the autocreated branch.

autoBranchCreationConfig_pullRequestEnvironmentName :: Lens' AutoBranchCreationConfig (Maybe Text)

The Amplify environment name for the pull request.