Amazonka.Amplify.Types.DomainAssociation

Description

Documentation

data DomainAssociation

Describes a domain association that associates a custom domain with an Amplify app.

See: newDomainAssociation smart constructor.

Constructors

DomainAssociation'

Fields

certificateVerificationDNSRecord :: Maybe Text

The DNS record for certificate verification.

autoSubDomainCreationPatterns :: Maybe [Text]

Sets branch patterns for automatic subdomain creation.

autoSubDomainIAMRole :: Maybe Text

The required AWS Identity and Access Management (IAM) service role for the Amazon Resource Name (ARN) for automatically creating subdomains.

domainAssociationArn :: Text

The Amazon Resource Name (ARN) for the domain association.

domainName :: Text

The name of the domain.

enableAutoSubDomain :: Bool

Enables the automated creation of subdomains for branches.

domainStatus :: DomainStatus

The current status of the domain association.

statusReason :: Text

The reason for the current status of the domain association.

subDomains :: [SubDomain]

The subdomains for the domain association.

newDomainAssociation

Arguments

:: Text
$sel:domainAssociationArn:DomainAssociation'

-> Text
$sel:domainName:DomainAssociation'

-> Bool
$sel:enableAutoSubDomain:DomainAssociation'

-> DomainStatus
$sel:domainStatus:DomainAssociation'

-> Text
$sel:statusReason:DomainAssociation'

-> DomainAssociation

Create a value of DomainAssociation with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:certificateVerificationDNSRecord:DomainAssociation', domainAssociation_certificateVerificationDNSRecord - The DNS record for certificate verification.

$sel:autoSubDomainCreationPatterns:DomainAssociation', domainAssociation_autoSubDomainCreationPatterns - Sets branch patterns for automatic subdomain creation.

$sel:autoSubDomainIAMRole:DomainAssociation', domainAssociation_autoSubDomainIAMRole - The required AWS Identity and Access Management (IAM) service role for the Amazon Resource Name (ARN) for automatically creating subdomains.

$sel:domainAssociationArn:DomainAssociation', domainAssociation_domainAssociationArn - The Amazon Resource Name (ARN) for the domain association.

$sel:domainName:DomainAssociation', domainAssociation_domainName - The name of the domain.

$sel:enableAutoSubDomain:DomainAssociation', domainAssociation_enableAutoSubDomain - Enables the automated creation of subdomains for branches.

$sel:domainStatus:DomainAssociation', domainAssociation_domainStatus - The current status of the domain association.

$sel:statusReason:DomainAssociation', domainAssociation_statusReason - The reason for the current status of the domain association.

$sel:subDomains:DomainAssociation', domainAssociation_subDomains - The subdomains for the domain association.

domainAssociation_certificateVerificationDNSRecord :: Lens' DomainAssociation (Maybe Text)

The DNS record for certificate verification.

domainAssociation_autoSubDomainCreationPatterns :: Lens' DomainAssociation (Maybe [Text])

Sets branch patterns for automatic subdomain creation.

domainAssociation_autoSubDomainIAMRole :: Lens' DomainAssociation (Maybe Text)

The required AWS Identity and Access Management (IAM) service role for the Amazon Resource Name (ARN) for automatically creating subdomains.

domainAssociation_domainAssociationArn :: Lens' DomainAssociation Text

The Amazon Resource Name (ARN) for the domain association.

domainAssociation_domainName :: Lens' DomainAssociation Text

The name of the domain.

domainAssociation_enableAutoSubDomain :: Lens' DomainAssociation Bool

Enables the automated creation of subdomains for branches.

domainAssociation_domainStatus :: Lens' DomainAssociation DomainStatus

The current status of the domain association.

domainAssociation_statusReason :: Lens' DomainAssociation Text

The reason for the current status of the domain association.

domainAssociation_subDomains :: Lens' DomainAssociation [SubDomain]

The subdomains for the domain association.