Amazonka.Amplify.Types.ProductionBranch

Description

Documentation

data ProductionBranch

Describes the information about a production branch for an Amplify app.

See: newProductionBranch smart constructor.

Constructors

ProductionBranch'

Fields

lastDeployTime :: Maybe POSIX

The last deploy time of the production branch.

status :: Maybe Text

The status of the production branch.

thumbnailUrl :: Maybe Text

The thumbnail URL for the production branch.

branchName :: Maybe Text

The branch name for the production branch.

newProductionBranch :: ProductionBranch

Create a value of ProductionBranch with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:lastDeployTime:ProductionBranch', productionBranch_lastDeployTime - The last deploy time of the production branch.

$sel:status:ProductionBranch', productionBranch_status - The status of the production branch.

$sel:thumbnailUrl:ProductionBranch', productionBranch_thumbnailUrl - The thumbnail URL for the production branch.

$sel:branchName:ProductionBranch', productionBranch_branchName - The branch name for the production branch.

productionBranch_lastDeployTime :: Lens' ProductionBranch (Maybe UTCTime)

The last deploy time of the production branch.

productionBranch_status :: Lens' ProductionBranch (Maybe Text)

The status of the production branch.

productionBranch_thumbnailUrl :: Lens' ProductionBranch (Maybe Text)

The thumbnail URL for the production branch.

productionBranch_branchName :: Lens' ProductionBranch (Maybe Text)

The branch name for the production branch.