Amazonka.Amplify.Types.Step

Description

Documentation

data Step

Describes an execution step, for an execution job, for an Amplify app.

See: newStep smart constructor.

Constructors

Step'

Fields

logUrl :: Maybe Text

The URL to the logs for the execution step.

context :: Maybe Text

The context for the current step. Includes a build image if the step is build.

testArtifactsUrl :: Maybe Text

The URL to the test artifact for the execution step.

artifactsUrl :: Maybe Text

The URL to the artifact for the execution step.

testConfigUrl :: Maybe Text

The URL to the test configuration for the execution step.

screenshots :: Maybe (HashMap Text Text)

The list of screenshot URLs for the execution step, if relevant.

statusReason :: Maybe Text

The reason for the current step status.

stepName :: Text

The name of the execution step.

startTime :: POSIX

The start date and time of the execution step.

status :: JobStatus

The status of the execution step.

endTime :: POSIX

The end date and time of the execution step.

newStep

Arguments

:: Text
$sel:stepName:Step'

-> UTCTime
$sel:startTime:Step'

-> JobStatus
$sel:status:Step'

-> UTCTime
$sel:endTime:Step'

-> Step

Create a value of Step with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:logUrl:Step', step_logUrl - The URL to the logs for the execution step.

$sel:context:Step', step_context - The context for the current step. Includes a build image if the step is build.

$sel:testArtifactsUrl:Step', step_testArtifactsUrl - The URL to the test artifact for the execution step.

$sel:artifactsUrl:Step', step_artifactsUrl - The URL to the artifact for the execution step.

$sel:testConfigUrl:Step', step_testConfigUrl - The URL to the test configuration for the execution step.

$sel:screenshots:Step', step_screenshots - The list of screenshot URLs for the execution step, if relevant.

$sel:statusReason:Step', step_statusReason - The reason for the current step status.

$sel:stepName:Step', step_stepName - The name of the execution step.

$sel:startTime:Step', step_startTime - The start date and time of the execution step.

$sel:status:Step', step_status - The status of the execution step.

$sel:endTime:Step', step_endTime - The end date and time of the execution step.

step_logUrl :: Lens' Step (Maybe Text)

The URL to the logs for the execution step.

step_context :: Lens' Step (Maybe Text)

The context for the current step. Includes a build image if the step is build.

step_testArtifactsUrl :: Lens' Step (Maybe Text)

The URL to the test artifact for the execution step.

step_artifactsUrl :: Lens' Step (Maybe Text)

The URL to the artifact for the execution step.

step_testConfigUrl :: Lens' Step (Maybe Text)

The URL to the test configuration for the execution step.

step_screenshots :: Lens' Step (Maybe (HashMap Text Text))

The list of screenshot URLs for the execution step, if relevant.

step_statusReason :: Lens' Step (Maybe Text)

The reason for the current step status.

step_stepName :: Lens' Step Text

The name of the execution step.

step_startTime :: Lens' Step UTCTime

The start date and time of the execution step.

step_status :: Lens' Step JobStatus

The status of the execution step.

step_endTime :: Lens' Step UTCTime

The end date and time of the execution step.