Amazonka.AmplifyBackend.CloneBackend

Description

This operation clones an existing backend.

Creating a Request

data CloneBackend

The request body for CloneBackend.

See: newCloneBackend smart constructor.

Constructors

CloneBackend'

Fields

appId :: Text

The app ID.

backendEnvironmentName :: Text

The name of the backend environment.

targetEnvironmentName :: Text

The name of the destination backend environment to be created.

newCloneBackend

Arguments

:: Text
$sel:appId:CloneBackend'

-> Text
$sel:backendEnvironmentName:CloneBackend'

-> Text
$sel:targetEnvironmentName:CloneBackend'

-> CloneBackend

Create a value of CloneBackend with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:appId:CloneBackend', cloneBackend_appId - The app ID.

$sel:backendEnvironmentName:CloneBackend', cloneBackend_backendEnvironmentName - The name of the backend environment.

$sel:targetEnvironmentName:CloneBackend', cloneBackend_targetEnvironmentName - The name of the destination backend environment to be created.

Request Lenses

cloneBackend_appId :: Lens' CloneBackend Text

The app ID.

cloneBackend_backendEnvironmentName :: Lens' CloneBackend Text

The name of the backend environment.

cloneBackend_targetEnvironmentName :: Lens' CloneBackend Text

The name of the destination backend environment to be created.

Destructuring the Response

data CloneBackendResponse

See: newCloneBackendResponse smart constructor.

Constructors

CloneBackendResponse'

Fields

status :: Maybe Text

The current status of the request.

jobId :: Maybe Text

The ID for the job.

operation :: Maybe Text

The name of the operation.

error :: Maybe Text

If the request fails, this error is returned.

appId :: Maybe Text

The app ID.

backendEnvironmentName :: Maybe Text

The name of the backend environment.

httpStatus :: Int

The response's http status code.

newCloneBackendResponse

Arguments

:: Int
$sel:httpStatus:CloneBackendResponse'

-> CloneBackendResponse

Create a value of CloneBackendResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:status:CloneBackendResponse', cloneBackendResponse_status - The current status of the request.

$sel:jobId:CloneBackendResponse', cloneBackendResponse_jobId - The ID for the job.

$sel:operation:CloneBackendResponse', cloneBackendResponse_operation - The name of the operation.

$sel:error:CloneBackendResponse', cloneBackendResponse_error - If the request fails, this error is returned.

$sel:appId:CloneBackend', cloneBackendResponse_appId - The app ID.

$sel:backendEnvironmentName:CloneBackend', cloneBackendResponse_backendEnvironmentName - The name of the backend environment.

$sel:httpStatus:CloneBackendResponse', cloneBackendResponse_httpStatus - The response's http status code.

Response Lenses

cloneBackendResponse_status :: Lens' CloneBackendResponse (Maybe Text)

The current status of the request.

cloneBackendResponse_jobId :: Lens' CloneBackendResponse (Maybe Text)

The ID for the job.

cloneBackendResponse_operation :: Lens' CloneBackendResponse (Maybe Text)

The name of the operation.

cloneBackendResponse_error :: Lens' CloneBackendResponse (Maybe Text)

If the request fails, this error is returned.

cloneBackendResponse_appId :: Lens' CloneBackendResponse (Maybe Text)

The app ID.

cloneBackendResponse_backendEnvironmentName :: Lens' CloneBackendResponse (Maybe Text)

The name of the backend environment.

cloneBackendResponse_httpStatus :: Lens' CloneBackendResponse Int

The response's http status code.