Amazonka.AmplifyBackend.CreateBackend

Description

This operation creates a backend for an Amplify app. Backends are automatically created at the time of app creation.

Creating a Request

data CreateBackend

The request body for CreateBackend.

See: newCreateBackend smart constructor.

Constructors

CreateBackend'

Fields

resourceName :: Maybe Text

The name of the resource.

resourceConfig :: Maybe ResourceConfig

The resource configuration for the create backend request.

appId :: Text

The app ID.

backendEnvironmentName :: Text

The name of the backend environment.

appName :: Text

The name of the app.

newCreateBackend

Arguments

:: Text
$sel:appId:CreateBackend'

-> Text
$sel:backendEnvironmentName:CreateBackend'

-> Text
$sel:appName:CreateBackend'

-> CreateBackend

Create a value of CreateBackend with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:resourceName:CreateBackend', createBackend_resourceName - The name of the resource.

$sel:resourceConfig:CreateBackend', createBackend_resourceConfig - The resource configuration for the create backend request.

$sel:appId:CreateBackend', createBackend_appId - The app ID.

$sel:backendEnvironmentName:CreateBackend', createBackend_backendEnvironmentName - The name of the backend environment.

$sel:appName:CreateBackend', createBackend_appName - The name of the app.

Request Lenses

createBackend_resourceName :: Lens' CreateBackend (Maybe Text)

The name of the resource.

createBackend_resourceConfig :: Lens' CreateBackend (Maybe ResourceConfig)

The resource configuration for the create backend request.

createBackend_appId :: Lens' CreateBackend Text

The app ID.

createBackend_backendEnvironmentName :: Lens' CreateBackend Text

The name of the backend environment.

createBackend_appName :: Lens' CreateBackend Text

The name of the app.

Destructuring the Response

data CreateBackendResponse

See: newCreateBackendResponse smart constructor.

Constructors

CreateBackendResponse'

Fields

status :: Maybe Text

The current status of the request.

jobId :: Maybe Text

The ID for the job.

operation :: Maybe Text

The name of the operation.

error :: Maybe Text

If the request fails, this error is returned.

appId :: Maybe Text

The app ID.

backendEnvironmentName :: Maybe Text

The name of the backend environment.

httpStatus :: Int

The response's http status code.

newCreateBackendResponse

Arguments

:: Int
$sel:httpStatus:CreateBackendResponse'

-> CreateBackendResponse

Create a value of CreateBackendResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:status:CreateBackendResponse', createBackendResponse_status - The current status of the request.

$sel:jobId:CreateBackendResponse', createBackendResponse_jobId - The ID for the job.

$sel:operation:CreateBackendResponse', createBackendResponse_operation - The name of the operation.

$sel:error:CreateBackendResponse', createBackendResponse_error - If the request fails, this error is returned.

$sel:appId:CreateBackend', createBackendResponse_appId - The app ID.

$sel:backendEnvironmentName:CreateBackend', createBackendResponse_backendEnvironmentName - The name of the backend environment.

$sel:httpStatus:CreateBackendResponse', createBackendResponse_httpStatus - The response's http status code.

Response Lenses

createBackendResponse_status :: Lens' CreateBackendResponse (Maybe Text)

The current status of the request.

createBackendResponse_jobId :: Lens' CreateBackendResponse (Maybe Text)

The ID for the job.

createBackendResponse_operation :: Lens' CreateBackendResponse (Maybe Text)

The name of the operation.

createBackendResponse_error :: Lens' CreateBackendResponse (Maybe Text)

If the request fails, this error is returned.

createBackendResponse_appId :: Lens' CreateBackendResponse (Maybe Text)

The app ID.

createBackendResponse_backendEnvironmentName :: Lens' CreateBackendResponse (Maybe Text)

The name of the backend environment.

createBackendResponse_httpStatus :: Lens' CreateBackendResponse Int

The response's http status code.