Amazonka.AmplifyBackend.CreateBackendConfig

Description

Creates a config object for a backend.

Creating a Request

data CreateBackendConfig

The request body for CreateBackendConfig.

See: newCreateBackendConfig smart constructor.

Constructors

CreateBackendConfig'

Fields

backendManagerAppId :: Maybe Text

The app ID for the backend manager.

appId :: Text

The app ID.

newCreateBackendConfig

Arguments

:: Text
$sel:appId:CreateBackendConfig'

-> CreateBackendConfig

Create a value of CreateBackendConfig with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:backendManagerAppId:CreateBackendConfig', createBackendConfig_backendManagerAppId - The app ID for the backend manager.

$sel:appId:CreateBackendConfig', createBackendConfig_appId - The app ID.

Request Lenses

createBackendConfig_backendManagerAppId :: Lens' CreateBackendConfig (Maybe Text)

The app ID for the backend manager.

createBackendConfig_appId :: Lens' CreateBackendConfig Text

The app ID.

Destructuring the Response

data CreateBackendConfigResponse

See: newCreateBackendConfigResponse smart constructor.

Constructors

CreateBackendConfigResponse'

Fields

status :: Maybe Text

The current status of the request.

jobId :: Maybe Text

The ID for the job.

appId :: Maybe Text

The app ID.

backendEnvironmentName :: Maybe Text

The name of the backend environment.

httpStatus :: Int

The response's http status code.

newCreateBackendConfigResponse

Arguments

:: Int
$sel:httpStatus:CreateBackendConfigResponse'

-> CreateBackendConfigResponse

Create a value of CreateBackendConfigResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:status:CreateBackendConfigResponse', createBackendConfigResponse_status - The current status of the request.

$sel:jobId:CreateBackendConfigResponse', createBackendConfigResponse_jobId - The ID for the job.

$sel:appId:CreateBackendConfig', createBackendConfigResponse_appId - The app ID.

$sel:backendEnvironmentName:CreateBackendConfigResponse', createBackendConfigResponse_backendEnvironmentName - The name of the backend environment.

$sel:httpStatus:CreateBackendConfigResponse', createBackendConfigResponse_httpStatus - The response's http status code.

Response Lenses

createBackendConfigResponse_status :: Lens' CreateBackendConfigResponse (Maybe Text)

The current status of the request.

createBackendConfigResponse_jobId :: Lens' CreateBackendConfigResponse (Maybe Text)

The ID for the job.

createBackendConfigResponse_appId :: Lens' CreateBackendConfigResponse (Maybe Text)

The app ID.

createBackendConfigResponse_backendEnvironmentName :: Lens' CreateBackendConfigResponse (Maybe Text)

The name of the backend environment.

createBackendConfigResponse_httpStatus :: Lens' CreateBackendConfigResponse Int

The response's http status code.