Amazonka.AmplifyBackend.CreateToken

Description

Generates a one-time challenge code to authenticate a user into your Amplify Admin UI.

Creating a Request

data CreateToken

See: newCreateToken smart constructor.

Constructors

CreateToken'

Fields

appId :: Text

The app ID.

newCreateToken

Arguments

:: Text
$sel:appId:CreateToken'

-> CreateToken

Create a value of CreateToken with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:appId:CreateToken', createToken_appId - The app ID.

Request Lenses

createToken_appId :: Lens' CreateToken Text

The app ID.

Destructuring the Response

data CreateTokenResponse

See: newCreateTokenResponse smart constructor.

Constructors

CreateTokenResponse'

Fields

ttl :: Maybe Text

The expiry time for the one-time generated token code.

appId :: Maybe Text

The app ID.

sessionId :: Maybe Text

A unique ID provided when creating a new challenge token.

challengeCode :: Maybe Text

One-time challenge code for authenticating into the Amplify Admin UI.

httpStatus :: Int

The response's http status code.

newCreateTokenResponse

Arguments

:: Int
$sel:httpStatus:CreateTokenResponse'

-> CreateTokenResponse

Create a value of CreateTokenResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:ttl:CreateTokenResponse', createTokenResponse_ttl - The expiry time for the one-time generated token code.

$sel:appId:CreateToken', createTokenResponse_appId - The app ID.

$sel:sessionId:CreateTokenResponse', createTokenResponse_sessionId - A unique ID provided when creating a new challenge token.

$sel:challengeCode:CreateTokenResponse', createTokenResponse_challengeCode - One-time challenge code for authenticating into the Amplify Admin UI.

$sel:httpStatus:CreateTokenResponse', createTokenResponse_httpStatus - The response's http status code.

Response Lenses

createTokenResponse_ttl :: Lens' CreateTokenResponse (Maybe Text)

The expiry time for the one-time generated token code.

createTokenResponse_appId :: Lens' CreateTokenResponse (Maybe Text)

The app ID.

createTokenResponse_sessionId :: Lens' CreateTokenResponse (Maybe Text)

A unique ID provided when creating a new challenge token.

createTokenResponse_challengeCode :: Lens' CreateTokenResponse (Maybe Text)

One-time challenge code for authenticating into the Amplify Admin UI.

createTokenResponse_httpStatus :: Lens' CreateTokenResponse Int

The response's http status code.