Amazonka.AmplifyBackend.GetBackendJob

Description

Returns information about a specific job.

Creating a Request

data GetBackendJob

See: newGetBackendJob smart constructor.

Constructors

GetBackendJob'

Fields

appId :: Text

The app ID.

backendEnvironmentName :: Text

The name of the backend environment.

jobId :: Text

The ID for the job.

newGetBackendJob

Arguments

:: Text
$sel:appId:GetBackendJob'

-> Text
$sel:backendEnvironmentName:GetBackendJob'

-> Text
$sel:jobId:GetBackendJob'

-> GetBackendJob

Create a value of GetBackendJob with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:appId:GetBackendJob', getBackendJob_appId - The app ID.

$sel:backendEnvironmentName:GetBackendJob', getBackendJob_backendEnvironmentName - The name of the backend environment.

$sel:jobId:GetBackendJob', getBackendJob_jobId - The ID for the job.

Request Lenses

getBackendJob_appId :: Lens' GetBackendJob Text

The app ID.

getBackendJob_backendEnvironmentName :: Lens' GetBackendJob Text

The name of the backend environment.

getBackendJob_jobId :: Lens' GetBackendJob Text

The ID for the job.

Destructuring the Response

data GetBackendJobResponse

See: newGetBackendJobResponse smart constructor.

Constructors

GetBackendJobResponse'

Fields

status :: Maybe Text

The current status of the request.

jobId :: Maybe Text

The ID for the job.

operation :: Maybe Text

The name of the operation.

error :: Maybe Text

If the request fails, this error is returned.

updateTime :: Maybe Text

The time when the job was last updated.

appId :: Maybe Text

The app ID.

backendEnvironmentName :: Maybe Text

The name of the backend environment.

createTime :: Maybe Text

The time when the job was created.

httpStatus :: Int

The response's http status code.

newGetBackendJobResponse

Arguments

:: Int
$sel:httpStatus:GetBackendJobResponse'

-> GetBackendJobResponse

Create a value of GetBackendJobResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:status:GetBackendJobResponse', getBackendJobResponse_status - The current status of the request.

$sel:jobId:GetBackendJob', getBackendJobResponse_jobId - The ID for the job.

$sel:operation:GetBackendJobResponse', getBackendJobResponse_operation - The name of the operation.

$sel:error:GetBackendJobResponse', getBackendJobResponse_error - If the request fails, this error is returned.

$sel:updateTime:GetBackendJobResponse', getBackendJobResponse_updateTime - The time when the job was last updated.

$sel:appId:GetBackendJob', getBackendJobResponse_appId - The app ID.

$sel:backendEnvironmentName:GetBackendJob', getBackendJobResponse_backendEnvironmentName - The name of the backend environment.

$sel:createTime:GetBackendJobResponse', getBackendJobResponse_createTime - The time when the job was created.

$sel:httpStatus:GetBackendJobResponse', getBackendJobResponse_httpStatus - The response's http status code.

Response Lenses

getBackendJobResponse_status :: Lens' GetBackendJobResponse (Maybe Text)

The current status of the request.

getBackendJobResponse_jobId :: Lens' GetBackendJobResponse (Maybe Text)

The ID for the job.

getBackendJobResponse_operation :: Lens' GetBackendJobResponse (Maybe Text)

The name of the operation.

getBackendJobResponse_error :: Lens' GetBackendJobResponse (Maybe Text)

If the request fails, this error is returned.

getBackendJobResponse_updateTime :: Lens' GetBackendJobResponse (Maybe Text)

The time when the job was last updated.

getBackendJobResponse_appId :: Lens' GetBackendJobResponse (Maybe Text)

The app ID.

getBackendJobResponse_backendEnvironmentName :: Lens' GetBackendJobResponse (Maybe Text)

The name of the backend environment.

getBackendJobResponse_createTime :: Lens' GetBackendJobResponse (Maybe Text)

The time when the job was created.

getBackendJobResponse_httpStatus :: Lens' GetBackendJobResponse Int

The response's http status code.