Amazonka.AmplifyBackend.GetToken

Description

Gets the challenge token based on the given appId and sessionId.

Creating a Request

data GetToken

See: newGetToken smart constructor.

Constructors

GetToken'

Fields

sessionId :: Text

The session ID.

appId :: Text

The app ID.

newGetToken

Arguments

:: Text
$sel:sessionId:GetToken'

-> Text
$sel:appId:GetToken'

-> GetToken

Create a value of GetToken with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:sessionId:GetToken', getToken_sessionId - The session ID.

$sel:appId:GetToken', getToken_appId - The app ID.

Request Lenses

getToken_sessionId :: Lens' GetToken Text

The session ID.

getToken_appId :: Lens' GetToken Text

The app ID.

Destructuring the Response

data GetTokenResponse

See: newGetTokenResponse smart constructor.

Constructors

GetTokenResponse'

Fields

ttl :: Maybe Text

The expiry time for the one-time generated token code.

appId :: Maybe Text

The app ID.

sessionId :: Maybe Text

A unique ID provided when creating a new challenge token.

challengeCode :: Maybe Text

The one-time challenge code for authenticating into the Amplify Admin UI.

httpStatus :: Int

The response's http status code.

newGetTokenResponse

Arguments

:: Int
$sel:httpStatus:GetTokenResponse'

-> GetTokenResponse

Create a value of GetTokenResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:ttl:GetTokenResponse', getTokenResponse_ttl - The expiry time for the one-time generated token code.

$sel:appId:GetToken', getTokenResponse_appId - The app ID.

$sel:sessionId:GetToken', getTokenResponse_sessionId - A unique ID provided when creating a new challenge token.

$sel:challengeCode:GetTokenResponse', getTokenResponse_challengeCode - The one-time challenge code for authenticating into the Amplify Admin UI.

$sel:httpStatus:GetTokenResponse', getTokenResponse_httpStatus - The response's http status code.

Response Lenses

getTokenResponse_ttl :: Lens' GetTokenResponse (Maybe Text)

The expiry time for the one-time generated token code.

getTokenResponse_appId :: Lens' GetTokenResponse (Maybe Text)

The app ID.

getTokenResponse_sessionId :: Lens' GetTokenResponse (Maybe Text)

A unique ID provided when creating a new challenge token.

getTokenResponse_challengeCode :: Lens' GetTokenResponse (Maybe Text)

The one-time challenge code for authenticating into the Amplify Admin UI.

getTokenResponse_httpStatus :: Lens' GetTokenResponse Int

The response's http status code.