Amazonka.AmplifyBackend.ListBackendJobs

Description

Lists the jobs for the backend of an Amplify app.

This operation returns paginated results.

Creating a Request

data ListBackendJobs

The request body for ListBackendJobs.

See: newListBackendJobs smart constructor.

Constructors

ListBackendJobs'

Fields

status :: Maybe Text

Filters the list of response objects to include only those with the specified status.

jobId :: Maybe Text

The ID for the job.

operation :: Maybe Text

Filters the list of response objects to include only those with the specified operation name.

nextToken :: Maybe Text

The token for the next set of results.

maxResults :: Maybe Natural

The maximum number of results that you want in the response.

appId :: Text

The app ID.

backendEnvironmentName :: Text

The name of the backend environment.

newListBackendJobs

Arguments

:: Text
$sel:appId:ListBackendJobs'

-> Text
$sel:backendEnvironmentName:ListBackendJobs'

-> ListBackendJobs

Create a value of ListBackendJobs with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:status:ListBackendJobs', listBackendJobs_status - Filters the list of response objects to include only those with the specified status.

$sel:jobId:ListBackendJobs', listBackendJobs_jobId - The ID for the job.

$sel:operation:ListBackendJobs', listBackendJobs_operation - Filters the list of response objects to include only those with the specified operation name.

$sel:nextToken:ListBackendJobs', listBackendJobs_nextToken - The token for the next set of results.

$sel:maxResults:ListBackendJobs', listBackendJobs_maxResults - The maximum number of results that you want in the response.

$sel:appId:ListBackendJobs', listBackendJobs_appId - The app ID.

$sel:backendEnvironmentName:ListBackendJobs', listBackendJobs_backendEnvironmentName - The name of the backend environment.

Request Lenses

listBackendJobs_status :: Lens' ListBackendJobs (Maybe Text)

Filters the list of response objects to include only those with the specified status.

listBackendJobs_jobId :: Lens' ListBackendJobs (Maybe Text)

The ID for the job.

listBackendJobs_operation :: Lens' ListBackendJobs (Maybe Text)

Filters the list of response objects to include only those with the specified operation name.

listBackendJobs_nextToken :: Lens' ListBackendJobs (Maybe Text)

The token for the next set of results.

listBackendJobs_maxResults :: Lens' ListBackendJobs (Maybe Natural)

The maximum number of results that you want in the response.

listBackendJobs_appId :: Lens' ListBackendJobs Text

The app ID.

listBackendJobs_backendEnvironmentName :: Lens' ListBackendJobs Text

The name of the backend environment.

Destructuring the Response

data ListBackendJobsResponse

See: newListBackendJobsResponse smart constructor.

Constructors

ListBackendJobsResponse'

Fields

nextToken :: Maybe Text

The token for the next set of results.

jobs :: Maybe [BackendJobRespObj]

An array of jobs and their properties.

httpStatus :: Int

The response's http status code.

newListBackendJobsResponse

Arguments

:: Int
$sel:httpStatus:ListBackendJobsResponse'

-> ListBackendJobsResponse

Create a value of ListBackendJobsResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:nextToken:ListBackendJobs', listBackendJobsResponse_nextToken - The token for the next set of results.

$sel:jobs:ListBackendJobsResponse', listBackendJobsResponse_jobs - An array of jobs and their properties.

$sel:httpStatus:ListBackendJobsResponse', listBackendJobsResponse_httpStatus - The response's http status code.

Response Lenses

listBackendJobsResponse_nextToken :: Lens' ListBackendJobsResponse (Maybe Text)

The token for the next set of results.

listBackendJobsResponse_jobs :: Lens' ListBackendJobsResponse (Maybe [BackendJobRespObj])

An array of jobs and their properties.

listBackendJobsResponse_httpStatus :: Lens' ListBackendJobsResponse Int

The response's http status code.