Amazonka.AmplifyBackend.Types.CreateBackendAuthForgotPasswordConfig

Description

Documentation

data CreateBackendAuthForgotPasswordConfig

Describes the forgot password policy for authenticating into the Amplify app.

See: newCreateBackendAuthForgotPasswordConfig smart constructor.

Constructors

CreateBackendAuthForgotPasswordConfig'

Fields

emailSettings :: Maybe EmailSettings

The configuration for the email sent when an app user forgets their password.

smsSettings :: Maybe SmsSettings

The configuration for the SMS message sent when an app user forgets their password.

deliveryMethod :: DeliveryMethod

Describes which mode to use (either SMS or email) to deliver messages to app users who want to recover their password.

newCreateBackendAuthForgotPasswordConfig

Arguments

:: DeliveryMethod
$sel:deliveryMethod:CreateBackendAuthForgotPasswordConfig'

-> CreateBackendAuthForgotPasswordConfig

Create a value of CreateBackendAuthForgotPasswordConfig with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:emailSettings:CreateBackendAuthForgotPasswordConfig', createBackendAuthForgotPasswordConfig_emailSettings - The configuration for the email sent when an app user forgets their password.

$sel:smsSettings:CreateBackendAuthForgotPasswordConfig', createBackendAuthForgotPasswordConfig_smsSettings - The configuration for the SMS message sent when an app user forgets their password.

$sel:deliveryMethod:CreateBackendAuthForgotPasswordConfig', createBackendAuthForgotPasswordConfig_deliveryMethod - Describes which mode to use (either SMS or email) to deliver messages to app users who want to recover their password.

createBackendAuthForgotPasswordConfig_emailSettings :: Lens' CreateBackendAuthForgotPasswordConfig (Maybe EmailSettings)

The configuration for the email sent when an app user forgets their password.

createBackendAuthForgotPasswordConfig_smsSettings :: Lens' CreateBackendAuthForgotPasswordConfig (Maybe SmsSettings)

The configuration for the SMS message sent when an app user forgets their password.

createBackendAuthForgotPasswordConfig_deliveryMethod :: Lens' CreateBackendAuthForgotPasswordConfig DeliveryMethod

Describes which mode to use (either SMS or email) to deliver messages to app users who want to recover their password.