Amazonka.AmplifyBackend.Types.SocialProviderSettings

Description

Documentation

data SocialProviderSettings

The settings for using the social identity providers for access to your Amplify app.

See: newSocialProviderSettings smart constructor.

Constructors

SocialProviderSettings'

Fields

loginWithAmazon :: Maybe BackendAuthSocialProviderConfig

signInWithApple :: Maybe BackendAuthAppleProviderConfig

facebook :: Maybe BackendAuthSocialProviderConfig

google :: Maybe BackendAuthSocialProviderConfig

newSocialProviderSettings :: SocialProviderSettings

Create a value of SocialProviderSettings with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:loginWithAmazon:SocialProviderSettings', socialProviderSettings_loginWithAmazon - Undocumented member.

$sel:signInWithApple:SocialProviderSettings', socialProviderSettings_signInWithApple - Undocumented member.

$sel:facebook:SocialProviderSettings', socialProviderSettings_facebook - Undocumented member.

$sel:google:SocialProviderSettings', socialProviderSettings_google - Undocumented member.

socialProviderSettings_loginWithAmazon :: Lens' SocialProviderSettings (Maybe BackendAuthSocialProviderConfig)

Undocumented member.

socialProviderSettings_signInWithApple :: Lens' SocialProviderSettings (Maybe BackendAuthAppleProviderConfig)

Undocumented member.

socialProviderSettings_facebook :: Lens' SocialProviderSettings (Maybe BackendAuthSocialProviderConfig)

Undocumented member.

socialProviderSettings_google :: Lens' SocialProviderSettings (Maybe BackendAuthSocialProviderConfig)

Undocumented member.