Amazonka.AmplifyBackend.Types.UpdateBackendAuthUserPoolConfig

Description

Documentation

data UpdateBackendAuthUserPoolConfig

Describes the Amazon Cognito user pool configuration for the authorization resource to be configured for your Amplify project on an update.

See: newUpdateBackendAuthUserPoolConfig smart constructor.

Constructors

UpdateBackendAuthUserPoolConfig'

Fields

passwordPolicy :: Maybe UpdateBackendAuthPasswordPolicyConfig

Describes the password policy for your Amazon Cognito user pool, configured as a part of your Amplify project.

mfa :: Maybe UpdateBackendAuthMFAConfig

Describes whether to apply multi-factor authentication policies for your Amazon Cognito user pool configured as a part of your Amplify project.

forgotPassword :: Maybe UpdateBackendAuthForgotPasswordConfig

Describes the forgot password policy for your Amazon Cognito user pool, configured as a part of your Amplify project.

oAuth :: Maybe UpdateBackendAuthOAuthConfig

Describes the OAuth policy and rules for your Amazon Cognito user pool, configured as a part of your Amplify project.

newUpdateBackendAuthUserPoolConfig :: UpdateBackendAuthUserPoolConfig

Create a value of UpdateBackendAuthUserPoolConfig with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:passwordPolicy:UpdateBackendAuthUserPoolConfig', updateBackendAuthUserPoolConfig_passwordPolicy - Describes the password policy for your Amazon Cognito user pool, configured as a part of your Amplify project.

$sel:mfa:UpdateBackendAuthUserPoolConfig', updateBackendAuthUserPoolConfig_mfa - Describes whether to apply multi-factor authentication policies for your Amazon Cognito user pool configured as a part of your Amplify project.

$sel:forgotPassword:UpdateBackendAuthUserPoolConfig', updateBackendAuthUserPoolConfig_forgotPassword - Describes the forgot password policy for your Amazon Cognito user pool, configured as a part of your Amplify project.

$sel:oAuth:UpdateBackendAuthUserPoolConfig', updateBackendAuthUserPoolConfig_oAuth - Describes the OAuth policy and rules for your Amazon Cognito user pool, configured as a part of your Amplify project.

updateBackendAuthUserPoolConfig_passwordPolicy :: Lens' UpdateBackendAuthUserPoolConfig (Maybe UpdateBackendAuthPasswordPolicyConfig)

Describes the password policy for your Amazon Cognito user pool, configured as a part of your Amplify project.

updateBackendAuthUserPoolConfig_mfa :: Lens' UpdateBackendAuthUserPoolConfig (Maybe UpdateBackendAuthMFAConfig)

Describes whether to apply multi-factor authentication policies for your Amazon Cognito user pool configured as a part of your Amplify project.

updateBackendAuthUserPoolConfig_forgotPassword :: Lens' UpdateBackendAuthUserPoolConfig (Maybe UpdateBackendAuthForgotPasswordConfig)

Describes the forgot password policy for your Amazon Cognito user pool, configured as a part of your Amplify project.

updateBackendAuthUserPoolConfig_oAuth :: Lens' UpdateBackendAuthUserPoolConfig (Maybe UpdateBackendAuthOAuthConfig)

Describes the OAuth policy and rules for your Amazon Cognito user pool, configured as a part of your Amplify project.