Amazonka.AmplifyBackend.UpdateBackendJob

Description

Updates a specific job.

Creating a Request

data UpdateBackendJob

The request body for GetBackendJob.

See: newUpdateBackendJob smart constructor.

Constructors

UpdateBackendJob'

Fields

status :: Maybe Text

Filters the list of response objects to include only those with the specified status.

operation :: Maybe Text

Filters the list of response objects to include only those with the specified operation name.

appId :: Text

The app ID.

backendEnvironmentName :: Text

The name of the backend environment.

jobId :: Text

The ID for the job.

newUpdateBackendJob

Arguments

:: Text
$sel:appId:UpdateBackendJob'

-> Text
$sel:backendEnvironmentName:UpdateBackendJob'

-> Text
$sel:jobId:UpdateBackendJob'

-> UpdateBackendJob

Create a value of UpdateBackendJob with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:status:UpdateBackendJob', updateBackendJob_status - Filters the list of response objects to include only those with the specified status.

$sel:operation:UpdateBackendJob', updateBackendJob_operation - Filters the list of response objects to include only those with the specified operation name.

$sel:appId:UpdateBackendJob', updateBackendJob_appId - The app ID.

$sel:backendEnvironmentName:UpdateBackendJob', updateBackendJob_backendEnvironmentName - The name of the backend environment.

$sel:jobId:UpdateBackendJob', updateBackendJob_jobId - The ID for the job.

Request Lenses

updateBackendJob_status :: Lens' UpdateBackendJob (Maybe Text)

Filters the list of response objects to include only those with the specified status.

updateBackendJob_operation :: Lens' UpdateBackendJob (Maybe Text)

Filters the list of response objects to include only those with the specified operation name.

updateBackendJob_appId :: Lens' UpdateBackendJob Text

The app ID.

updateBackendJob_backendEnvironmentName :: Lens' UpdateBackendJob Text

The name of the backend environment.

updateBackendJob_jobId :: Lens' UpdateBackendJob Text

The ID for the job.

Destructuring the Response

data UpdateBackendJobResponse

See: newUpdateBackendJobResponse smart constructor.

Constructors

UpdateBackendJobResponse'

Fields

status :: Maybe Text

The current status of the request.

jobId :: Maybe Text

The ID for the job.

operation :: Maybe Text

The name of the operation.

error :: Maybe Text

If the request fails, this error is returned.

updateTime :: Maybe Text

The time when the job was last updated.

appId :: Maybe Text

The app ID.

backendEnvironmentName :: Maybe Text

The name of the backend environment.

createTime :: Maybe Text

The time when the job was created.

httpStatus :: Int

The response's http status code.

newUpdateBackendJobResponse

Arguments

:: Int
$sel:httpStatus:UpdateBackendJobResponse'

-> UpdateBackendJobResponse

Create a value of UpdateBackendJobResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:status:UpdateBackendJob', updateBackendJobResponse_status - The current status of the request.

$sel:jobId:UpdateBackendJob', updateBackendJobResponse_jobId - The ID for the job.

$sel:operation:UpdateBackendJob', updateBackendJobResponse_operation - The name of the operation.

$sel:error:UpdateBackendJobResponse', updateBackendJobResponse_error - If the request fails, this error is returned.

$sel:updateTime:UpdateBackendJobResponse', updateBackendJobResponse_updateTime - The time when the job was last updated.

$sel:appId:UpdateBackendJob', updateBackendJobResponse_appId - The app ID.

$sel:backendEnvironmentName:UpdateBackendJob', updateBackendJobResponse_backendEnvironmentName - The name of the backend environment.

$sel:createTime:UpdateBackendJobResponse', updateBackendJobResponse_createTime - The time when the job was created.

$sel:httpStatus:UpdateBackendJobResponse', updateBackendJobResponse_httpStatus - The response's http status code.

Response Lenses

updateBackendJobResponse_status :: Lens' UpdateBackendJobResponse (Maybe Text)

The current status of the request.

updateBackendJobResponse_jobId :: Lens' UpdateBackendJobResponse (Maybe Text)

The ID for the job.

updateBackendJobResponse_operation :: Lens' UpdateBackendJobResponse (Maybe Text)

The name of the operation.

updateBackendJobResponse_error :: Lens' UpdateBackendJobResponse (Maybe Text)

If the request fails, this error is returned.

updateBackendJobResponse_updateTime :: Lens' UpdateBackendJobResponse (Maybe Text)

The time when the job was last updated.

updateBackendJobResponse_appId :: Lens' UpdateBackendJobResponse (Maybe Text)

The app ID.

updateBackendJobResponse_backendEnvironmentName :: Lens' UpdateBackendJobResponse (Maybe Text)

The name of the backend environment.

updateBackendJobResponse_createTime :: Lens' UpdateBackendJobResponse (Maybe Text)

The time when the job was created.

updateBackendJobResponse_httpStatus :: Lens' UpdateBackendJobResponse Int

The response's http status code.