Amazonka.ApiGatewayV2.CreateDeployment

Description

Creates a Deployment for an API.

Creating a Request

data CreateDeployment

Creates a new Deployment resource to represent a deployment.

See: newCreateDeployment smart constructor.

Constructors

CreateDeployment'

Fields

stageName :: Maybe Text

The name of the Stage resource for the Deployment resource to create.

description :: Maybe Text

The description for the deployment resource.

apiId :: Text

The API identifier.

newCreateDeployment

Arguments

:: Text
$sel:apiId:CreateDeployment'

-> CreateDeployment

Create a value of CreateDeployment with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:stageName:CreateDeployment', createDeployment_stageName - The name of the Stage resource for the Deployment resource to create.

$sel:description:CreateDeployment', createDeployment_description - The description for the deployment resource.

$sel:apiId:CreateDeployment', createDeployment_apiId - The API identifier.

Request Lenses

createDeployment_stageName :: Lens' CreateDeployment (Maybe Text)

The name of the Stage resource for the Deployment resource to create.

createDeployment_description :: Lens' CreateDeployment (Maybe Text)

The description for the deployment resource.

createDeployment_apiId :: Lens' CreateDeployment Text

The API identifier.

Destructuring the Response

data CreateDeploymentResponse

See: newCreateDeploymentResponse smart constructor.

Constructors

CreateDeploymentResponse'

Fields

deploymentId :: Maybe Text

The identifier for the deployment.

autoDeployed :: Maybe Bool

Specifies whether a deployment was automatically released.

deploymentStatusMessage :: Maybe Text

May contain additional feedback on the status of an API deployment.

createdDate :: Maybe POSIX

The date and time when the Deployment resource was created.

deploymentStatus :: Maybe DeploymentStatus

The status of the deployment: PENDING, FAILED, or SUCCEEDED.

description :: Maybe Text

The description for the deployment.

httpStatus :: Int

The response's http status code.

newCreateDeploymentResponse

Arguments

:: Int
$sel:httpStatus:CreateDeploymentResponse'

-> CreateDeploymentResponse

Create a value of CreateDeploymentResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:deploymentId:CreateDeploymentResponse', createDeploymentResponse_deploymentId - The identifier for the deployment.

$sel:autoDeployed:CreateDeploymentResponse', createDeploymentResponse_autoDeployed - Specifies whether a deployment was automatically released.

$sel:deploymentStatusMessage:CreateDeploymentResponse', createDeploymentResponse_deploymentStatusMessage - May contain additional feedback on the status of an API deployment.

$sel:createdDate:CreateDeploymentResponse', createDeploymentResponse_createdDate - The date and time when the Deployment resource was created.

$sel:deploymentStatus:CreateDeploymentResponse', createDeploymentResponse_deploymentStatus - The status of the deployment: PENDING, FAILED, or SUCCEEDED.

$sel:description:CreateDeployment', createDeploymentResponse_description - The description for the deployment.

$sel:httpStatus:CreateDeploymentResponse', createDeploymentResponse_httpStatus - The response's http status code.

Response Lenses

createDeploymentResponse_deploymentId :: Lens' CreateDeploymentResponse (Maybe Text)

The identifier for the deployment.

createDeploymentResponse_autoDeployed :: Lens' CreateDeploymentResponse (Maybe Bool)

Specifies whether a deployment was automatically released.

createDeploymentResponse_deploymentStatusMessage :: Lens' CreateDeploymentResponse (Maybe Text)

May contain additional feedback on the status of an API deployment.

createDeploymentResponse_createdDate :: Lens' CreateDeploymentResponse (Maybe UTCTime)

The date and time when the Deployment resource was created.

createDeploymentResponse_deploymentStatus :: Lens' CreateDeploymentResponse (Maybe DeploymentStatus)

The status of the deployment: PENDING, FAILED, or SUCCEEDED.

createDeploymentResponse_description :: Lens' CreateDeploymentResponse (Maybe Text)

The description for the deployment.

createDeploymentResponse_httpStatus :: Lens' CreateDeploymentResponse Int

The response's http status code.