Amazonka.ApiGatewayV2.CreateDomainName

Description

Creates a domain name.

Creating a Request

data CreateDomainName

Creates a new DomainName resource to represent a domain name.

See: newCreateDomainName smart constructor.

Constructors

CreateDomainName'

Fields

domainNameConfigurations :: Maybe [DomainNameConfiguration]

The domain name configurations.

mutualTlsAuthentication :: Maybe MutualTlsAuthenticationInput

The mutual TLS authentication configuration for a custom domain name.

tags :: Maybe (HashMap Text Text)

The collection of tags associated with a domain name.

domainName :: Text

The domain name.

newCreateDomainName

Arguments

:: Text
$sel:domainName:CreateDomainName'

-> CreateDomainName

Create a value of CreateDomainName with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:domainNameConfigurations:CreateDomainName', createDomainName_domainNameConfigurations - The domain name configurations.

$sel:mutualTlsAuthentication:CreateDomainName', createDomainName_mutualTlsAuthentication - The mutual TLS authentication configuration for a custom domain name.

$sel:tags:CreateDomainName', createDomainName_tags - The collection of tags associated with a domain name.

$sel:domainName:CreateDomainName', createDomainName_domainName - The domain name.

Request Lenses

createDomainName_domainNameConfigurations :: Lens' CreateDomainName (Maybe [DomainNameConfiguration])

The domain name configurations.

createDomainName_mutualTlsAuthentication :: Lens' CreateDomainName (Maybe MutualTlsAuthenticationInput)

The mutual TLS authentication configuration for a custom domain name.

createDomainName_tags :: Lens' CreateDomainName (Maybe (HashMap Text Text))

The collection of tags associated with a domain name.

createDomainName_domainName :: Lens' CreateDomainName Text

The domain name.

Destructuring the Response

data CreateDomainNameResponse

See: newCreateDomainNameResponse smart constructor.

Constructors

CreateDomainNameResponse'

Fields

domainNameConfigurations :: Maybe [DomainNameConfiguration]

The domain name configurations.

domainName :: Maybe Text

The name of the DomainName resource.

mutualTlsAuthentication :: Maybe MutualTlsAuthentication

The mutual TLS authentication configuration for a custom domain name.

apiMappingSelectionExpression :: Maybe Text

The API mapping selection expression.

tags :: Maybe (HashMap Text Text)

The collection of tags associated with a domain name.

httpStatus :: Int

The response's http status code.

newCreateDomainNameResponse

Arguments

:: Int
$sel:httpStatus:CreateDomainNameResponse'

-> CreateDomainNameResponse

Create a value of CreateDomainNameResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:domainNameConfigurations:CreateDomainName', createDomainNameResponse_domainNameConfigurations - The domain name configurations.

$sel:domainName:CreateDomainName', createDomainNameResponse_domainName - The name of the DomainName resource.

$sel:mutualTlsAuthentication:CreateDomainName', createDomainNameResponse_mutualTlsAuthentication - The mutual TLS authentication configuration for a custom domain name.

$sel:apiMappingSelectionExpression:CreateDomainNameResponse', createDomainNameResponse_apiMappingSelectionExpression - The API mapping selection expression.

$sel:tags:CreateDomainName', createDomainNameResponse_tags - The collection of tags associated with a domain name.

$sel:httpStatus:CreateDomainNameResponse', createDomainNameResponse_httpStatus - The response's http status code.

Response Lenses

createDomainNameResponse_domainNameConfigurations :: Lens' CreateDomainNameResponse (Maybe [DomainNameConfiguration])

The domain name configurations.

createDomainNameResponse_domainName :: Lens' CreateDomainNameResponse (Maybe Text)

The name of the DomainName resource.

createDomainNameResponse_mutualTlsAuthentication :: Lens' CreateDomainNameResponse (Maybe MutualTlsAuthentication)

The mutual TLS authentication configuration for a custom domain name.

createDomainNameResponse_apiMappingSelectionExpression :: Lens' CreateDomainNameResponse (Maybe Text)

The API mapping selection expression.

createDomainNameResponse_tags :: Lens' CreateDomainNameResponse (Maybe (HashMap Text Text))

The collection of tags associated with a domain name.

createDomainNameResponse_httpStatus :: Lens' CreateDomainNameResponse Int

The response's http status code.