Amazonka.ApiGatewayV2.CreateIntegrationResponse

Description

Creates an IntegrationResponses.

Creating a Request

data CreateIntegrationResponse

Creates a new IntegrationResponse resource to represent an integration response.

See: newCreateIntegrationResponse smart constructor.

Constructors

CreateIntegrationResponse'

Fields

templateSelectionExpression :: Maybe Text

The template selection expression for the integration response. Supported only for WebSocket APIs.

contentHandlingStrategy :: Maybe ContentHandlingStrategy

Specifies how to handle response payload content type conversions. Supported values are CONVERT_TO_BINARY and CONVERT_TO_TEXT, with the following behaviors:

CONVERT_TO_BINARY: Converts a response payload from a Base64-encoded string to the corresponding binary blob.

CONVERT_TO_TEXT: Converts a response payload from a binary blob to a Base64-encoded string.

If this property is not defined, the response payload will be passed through from the integration response to the route response or method response without modification.

responseTemplates :: Maybe (HashMap Text Text)

The collection of response templates for the integration response as a string-to-string map of key-value pairs. Response templates are represented as a key/value map, with a content-type as the key and a template as the value.

responseParameters :: Maybe (HashMap Text Text)

A key-value map specifying response parameters that are passed to the method response from the backend. The key is a method response header parameter name and the mapped value is an integration response header value, a static value enclosed within a pair of single quotes, or a JSON expression from the integration response body. The mapping key must match the pattern of method.response.header.{name}, where {name} is a valid and unique header name. The mapped non-static value must match the pattern of integration.response.header.{name} or integration.response.body.{JSON-expression}, where {name} is a valid and unique response header name and {JSON-expression} is a valid JSON expression without the $ prefix.

apiId :: Text

The API identifier.

integrationId :: Text

The integration ID.

integrationResponseKey :: Text

The integration response key.

newCreateIntegrationResponse

Arguments

:: Text
$sel:apiId:CreateIntegrationResponse'

-> Text
$sel:integrationId:CreateIntegrationResponse'

-> Text
$sel:integrationResponseKey:CreateIntegrationResponse'

-> CreateIntegrationResponse

Create a value of CreateIntegrationResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:templateSelectionExpression:CreateIntegrationResponse', createIntegrationResponse_templateSelectionExpression - The template selection expression for the integration response. Supported only for WebSocket APIs.

$sel:contentHandlingStrategy:CreateIntegrationResponse', createIntegrationResponse_contentHandlingStrategy - Specifies how to handle response payload content type conversions. Supported values are CONVERT_TO_BINARY and CONVERT_TO_TEXT, with the following behaviors:

CONVERT_TO_BINARY: Converts a response payload from a Base64-encoded string to the corresponding binary blob.

CONVERT_TO_TEXT: Converts a response payload from a binary blob to a Base64-encoded string.

If this property is not defined, the response payload will be passed through from the integration response to the route response or method response without modification.

$sel:responseTemplates:CreateIntegrationResponse', createIntegrationResponse_responseTemplates - The collection of response templates for the integration response as a string-to-string map of key-value pairs. Response templates are represented as a key/value map, with a content-type as the key and a template as the value.

$sel:responseParameters:CreateIntegrationResponse', createIntegrationResponse_responseParameters - A key-value map specifying response parameters that are passed to the method response from the backend. The key is a method response header parameter name and the mapped value is an integration response header value, a static value enclosed within a pair of single quotes, or a JSON expression from the integration response body. The mapping key must match the pattern of method.response.header.{name}, where {name} is a valid and unique header name. The mapped non-static value must match the pattern of integration.response.header.{name} or integration.response.body.{JSON-expression}, where {name} is a valid and unique response header name and {JSON-expression} is a valid JSON expression without the $ prefix.

$sel:apiId:CreateIntegrationResponse', createIntegrationResponse_apiId - The API identifier.

$sel:integrationId:CreateIntegrationResponse', createIntegrationResponse_integrationId - The integration ID.

$sel:integrationResponseKey:CreateIntegrationResponse', createIntegrationResponse_integrationResponseKey - The integration response key.

Request Lenses

createIntegrationResponse_templateSelectionExpression :: Lens' CreateIntegrationResponse (Maybe Text)

The template selection expression for the integration response. Supported only for WebSocket APIs.

createIntegrationResponse_contentHandlingStrategy :: Lens' CreateIntegrationResponse (Maybe ContentHandlingStrategy)

Specifies how to handle response payload content type conversions. Supported values are CONVERT_TO_BINARY and CONVERT_TO_TEXT, with the following behaviors:

CONVERT_TO_BINARY: Converts a response payload from a Base64-encoded string to the corresponding binary blob.

CONVERT_TO_TEXT: Converts a response payload from a binary blob to a Base64-encoded string.

If this property is not defined, the response payload will be passed through from the integration response to the route response or method response without modification.

createIntegrationResponse_responseTemplates :: Lens' CreateIntegrationResponse (Maybe (HashMap Text Text))

The collection of response templates for the integration response as a string-to-string map of key-value pairs. Response templates are represented as a key/value map, with a content-type as the key and a template as the value.

createIntegrationResponse_responseParameters :: Lens' CreateIntegrationResponse (Maybe (HashMap Text Text))

A key-value map specifying response parameters that are passed to the method response from the backend. The key is a method response header parameter name and the mapped value is an integration response header value, a static value enclosed within a pair of single quotes, or a JSON expression from the integration response body. The mapping key must match the pattern of method.response.header.{name}, where {name} is a valid and unique header name. The mapped non-static value must match the pattern of integration.response.header.{name} or integration.response.body.{JSON-expression}, where {name} is a valid and unique response header name and {JSON-expression} is a valid JSON expression without the $ prefix.

createIntegrationResponse_apiId :: Lens' CreateIntegrationResponse Text

The API identifier.

createIntegrationResponse_integrationId :: Lens' CreateIntegrationResponse Text

The integration ID.

createIntegrationResponse_integrationResponseKey :: Lens' CreateIntegrationResponse Text

The integration response key.

Destructuring the Response

data CreateIntegrationResponseResponse

See: newCreateIntegrationResponseResponse smart constructor.

Constructors

CreateIntegrationResponseResponse'

Fields

integrationResponseId :: Maybe Text

The integration response ID.

integrationResponseKey :: Maybe Text

The integration response key.

templateSelectionExpression :: Maybe Text

The template selection expressions for the integration response.

contentHandlingStrategy :: Maybe ContentHandlingStrategy

Supported only for WebSocket APIs. Specifies how to handle response payload content type conversions. Supported values are CONVERT_TO_BINARY and CONVERT_TO_TEXT, with the following behaviors:

CONVERT_TO_BINARY: Converts a response payload from a Base64-encoded string to the corresponding binary blob.

CONVERT_TO_TEXT: Converts a response payload from a binary blob to a Base64-encoded string.

If this property is not defined, the response payload will be passed through from the integration response to the route response or method response without modification.

responseTemplates :: Maybe (HashMap Text Text)

The collection of response templates for the integration response as a string-to-string map of key-value pairs. Response templates are represented as a key/value map, with a content-type as the key and a template as the value.

responseParameters :: Maybe (HashMap Text Text)

A key-value map specifying response parameters that are passed to the method response from the backend. The key is a method response header parameter name and the mapped value is an integration response header value, a static value enclosed within a pair of single quotes, or a JSON expression from the integration response body. The mapping key must match the pattern of method.response.header.{name}, where name is a valid and unique header name. The mapped non-static value must match the pattern of integration.response.header.{name} or integration.response.body.{JSON-expression}, where name is a valid and unique response header name and JSON-expression is a valid JSON expression without the $ prefix.

httpStatus :: Int

The response's http status code.

newCreateIntegrationResponseResponse

Arguments

:: Int
$sel:httpStatus:CreateIntegrationResponseResponse'

-> CreateIntegrationResponseResponse

Create a value of CreateIntegrationResponseResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:integrationResponseId:CreateIntegrationResponseResponse', createIntegrationResponseResponse_integrationResponseId - The integration response ID.

$sel:integrationResponseKey:CreateIntegrationResponse', createIntegrationResponseResponse_integrationResponseKey - The integration response key.

$sel:templateSelectionExpression:CreateIntegrationResponse', createIntegrationResponseResponse_templateSelectionExpression - The template selection expressions for the integration response.

$sel:contentHandlingStrategy:CreateIntegrationResponse', createIntegrationResponseResponse_contentHandlingStrategy - Supported only for WebSocket APIs. Specifies how to handle response payload content type conversions. Supported values are CONVERT_TO_BINARY and CONVERT_TO_TEXT, with the following behaviors:

CONVERT_TO_BINARY: Converts a response payload from a Base64-encoded string to the corresponding binary blob.

CONVERT_TO_TEXT: Converts a response payload from a binary blob to a Base64-encoded string.

If this property is not defined, the response payload will be passed through from the integration response to the route response or method response without modification.

$sel:responseTemplates:CreateIntegrationResponse', createIntegrationResponseResponse_responseTemplates - The collection of response templates for the integration response as a string-to-string map of key-value pairs. Response templates are represented as a key/value map, with a content-type as the key and a template as the value.

$sel:responseParameters:CreateIntegrationResponse', createIntegrationResponseResponse_responseParameters - A key-value map specifying response parameters that are passed to the method response from the backend. The key is a method response header parameter name and the mapped value is an integration response header value, a static value enclosed within a pair of single quotes, or a JSON expression from the integration response body. The mapping key must match the pattern of method.response.header.{name}, where name is a valid and unique header name. The mapped non-static value must match the pattern of integration.response.header.{name} or integration.response.body.{JSON-expression}, where name is a valid and unique response header name and JSON-expression is a valid JSON expression without the $ prefix.

$sel:httpStatus:CreateIntegrationResponseResponse', createIntegrationResponseResponse_httpStatus - The response's http status code.

Response Lenses

createIntegrationResponseResponse_integrationResponseId :: Lens' CreateIntegrationResponseResponse (Maybe Text)

The integration response ID.

createIntegrationResponseResponse_integrationResponseKey :: Lens' CreateIntegrationResponseResponse (Maybe Text)

The integration response key.

createIntegrationResponseResponse_templateSelectionExpression :: Lens' CreateIntegrationResponseResponse (Maybe Text)

The template selection expressions for the integration response.

createIntegrationResponseResponse_contentHandlingStrategy :: Lens' CreateIntegrationResponseResponse (Maybe ContentHandlingStrategy)

Supported only for WebSocket APIs. Specifies how to handle response payload content type conversions. Supported values are CONVERT_TO_BINARY and CONVERT_TO_TEXT, with the following behaviors:

CONVERT_TO_BINARY: Converts a response payload from a Base64-encoded string to the corresponding binary blob.

CONVERT_TO_TEXT: Converts a response payload from a binary blob to a Base64-encoded string.

If this property is not defined, the response payload will be passed through from the integration response to the route response or method response without modification.

createIntegrationResponseResponse_responseTemplates :: Lens' CreateIntegrationResponseResponse (Maybe (HashMap Text Text))

The collection of response templates for the integration response as a string-to-string map of key-value pairs. Response templates are represented as a key/value map, with a content-type as the key and a template as the value.

createIntegrationResponseResponse_responseParameters :: Lens' CreateIntegrationResponseResponse (Maybe (HashMap Text Text))

A key-value map specifying response parameters that are passed to the method response from the backend. The key is a method response header parameter name and the mapped value is an integration response header value, a static value enclosed within a pair of single quotes, or a JSON expression from the integration response body. The mapping key must match the pattern of method.response.header.{name}, where name is a valid and unique header name. The mapped non-static value must match the pattern of integration.response.header.{name} or integration.response.body.{JSON-expression}, where name is a valid and unique response header name and JSON-expression is a valid JSON expression without the $ prefix.

createIntegrationResponseResponse_httpStatus :: Lens' CreateIntegrationResponseResponse Int

The response's http status code.