Amazonka.ApiGatewayV2.CreateRoute

Description

Creates a Route for an API.

Creating a Request

data CreateRoute

Creates a new Route resource to represent a route.

See: newCreateRoute smart constructor.

Constructors

CreateRoute'

Fields

authorizationScopes :: Maybe [Text]

The authorization scopes supported by this route.

modelSelectionExpression :: Maybe Text

The model selection expression for the route. Supported only for WebSocket APIs.

requestModels :: Maybe (HashMap Text Text)

The request models for the route. Supported only for WebSocket APIs.

routeResponseSelectionExpression :: Maybe Text

The route response selection expression for the route. Supported only for WebSocket APIs.

requestParameters :: Maybe (HashMap Text ParameterConstraints)

The request parameters for the route. Supported only for WebSocket APIs.

authorizerId :: Maybe Text

The identifier of the Authorizer resource to be associated with this route. The authorizer identifier is generated by API Gateway when you created the authorizer.

operationName :: Maybe Text

The operation name for the route.

authorizationType :: Maybe AuthorizationType

The authorization type for the route. For WebSocket APIs, valid values are NONE for open access, AWS_IAM for using AWS IAM permissions, and CUSTOM for using a Lambda authorizer For HTTP APIs, valid values are NONE for open access, JWT for using JSON Web Tokens, AWS_IAM for using AWS IAM permissions, and CUSTOM for using a Lambda authorizer.

apiKeyRequired :: Maybe Bool

Specifies whether an API key is required for the route. Supported only for WebSocket APIs.

target :: Maybe Text

The target for the route.

apiId :: Text

The API identifier.

routeKey :: Text

The route key for the route.

newCreateRoute

Arguments

:: Text
$sel:apiId:CreateRoute'

-> Text
$sel:routeKey:CreateRoute'

-> CreateRoute

Create a value of CreateRoute with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:authorizationScopes:CreateRoute', createRoute_authorizationScopes - The authorization scopes supported by this route.

$sel:modelSelectionExpression:CreateRoute', createRoute_modelSelectionExpression - The model selection expression for the route. Supported only for WebSocket APIs.

$sel:requestModels:CreateRoute', createRoute_requestModels - The request models for the route. Supported only for WebSocket APIs.

$sel:routeResponseSelectionExpression:CreateRoute', createRoute_routeResponseSelectionExpression - The route response selection expression for the route. Supported only for WebSocket APIs.

$sel:requestParameters:CreateRoute', createRoute_requestParameters - The request parameters for the route. Supported only for WebSocket APIs.

$sel:authorizerId:CreateRoute', createRoute_authorizerId - The identifier of the Authorizer resource to be associated with this route. The authorizer identifier is generated by API Gateway when you created the authorizer.

$sel:operationName:CreateRoute', createRoute_operationName - The operation name for the route.

$sel:authorizationType:CreateRoute', createRoute_authorizationType - The authorization type for the route. For WebSocket APIs, valid values are NONE for open access, AWS_IAM for using AWS IAM permissions, and CUSTOM for using a Lambda authorizer For HTTP APIs, valid values are NONE for open access, JWT for using JSON Web Tokens, AWS_IAM for using AWS IAM permissions, and CUSTOM for using a Lambda authorizer.

$sel:apiKeyRequired:CreateRoute', createRoute_apiKeyRequired - Specifies whether an API key is required for the route. Supported only for WebSocket APIs.

$sel:target:CreateRoute', createRoute_target - The target for the route.

$sel:apiId:CreateRoute', createRoute_apiId - The API identifier.

$sel:routeKey:CreateRoute', createRoute_routeKey - The route key for the route.

Request Lenses

createRoute_authorizationScopes :: Lens' CreateRoute (Maybe [Text])

The authorization scopes supported by this route.

createRoute_modelSelectionExpression :: Lens' CreateRoute (Maybe Text)

The model selection expression for the route. Supported only for WebSocket APIs.

createRoute_requestModels :: Lens' CreateRoute (Maybe (HashMap Text Text))

The request models for the route. Supported only for WebSocket APIs.

createRoute_routeResponseSelectionExpression :: Lens' CreateRoute (Maybe Text)

The route response selection expression for the route. Supported only for WebSocket APIs.

createRoute_requestParameters :: Lens' CreateRoute (Maybe (HashMap Text ParameterConstraints))

The request parameters for the route. Supported only for WebSocket APIs.

createRoute_authorizerId :: Lens' CreateRoute (Maybe Text)

The identifier of the Authorizer resource to be associated with this route. The authorizer identifier is generated by API Gateway when you created the authorizer.

createRoute_operationName :: Lens' CreateRoute (Maybe Text)

The operation name for the route.

createRoute_authorizationType :: Lens' CreateRoute (Maybe AuthorizationType)

The authorization type for the route. For WebSocket APIs, valid values are NONE for open access, AWS_IAM for using AWS IAM permissions, and CUSTOM for using a Lambda authorizer For HTTP APIs, valid values are NONE for open access, JWT for using JSON Web Tokens, AWS_IAM for using AWS IAM permissions, and CUSTOM for using a Lambda authorizer.

createRoute_apiKeyRequired :: Lens' CreateRoute (Maybe Bool)

Specifies whether an API key is required for the route. Supported only for WebSocket APIs.

createRoute_target :: Lens' CreateRoute (Maybe Text)

The target for the route.

createRoute_apiId :: Lens' CreateRoute Text

The API identifier.

createRoute_routeKey :: Lens' CreateRoute Text

The route key for the route.

Destructuring the Response

data CreateRouteResponse'

See: newCreateRouteResponse' smart constructor.

Constructors

CreateRouteResponse''

Fields

authorizationScopes :: Maybe [Text]

A list of authorization scopes configured on a route. The scopes are used with a JWT authorizer to authorize the method invocation. The authorization works by matching the route scopes against the scopes parsed from the access token in the incoming request. The method invocation is authorized if any route scope matches a claimed scope in the access token. Otherwise, the invocation is not authorized. When the route scope is configured, the client must provide an access token instead of an identity token for authorization purposes.

modelSelectionExpression :: Maybe Text

The model selection expression for the route. Supported only for WebSocket APIs.

requestModels :: Maybe (HashMap Text Text)

The request models for the route. Supported only for WebSocket APIs.

routeResponseSelectionExpression :: Maybe Text

The route response selection expression for the route. Supported only for WebSocket APIs.

requestParameters :: Maybe (HashMap Text ParameterConstraints)

The request parameters for the route. Supported only for WebSocket APIs.

routeId :: Maybe Text

The route ID.

authorizerId :: Maybe Text

The identifier of the Authorizer resource to be associated with this route. The authorizer identifier is generated by API Gateway when you created the authorizer.

operationName :: Maybe Text

The operation name for the route.

apiGatewayManaged :: Maybe Bool

Specifies whether a route is managed by API Gateway. If you created an API using quick create, the $default route is managed by API Gateway. You can't modify the $default route key.

authorizationType :: Maybe AuthorizationType

The authorization type for the route. For WebSocket APIs, valid values are NONE for open access, AWS_IAM for using AWS IAM permissions, and CUSTOM for using a Lambda authorizer For HTTP APIs, valid values are NONE for open access, JWT for using JSON Web Tokens, AWS_IAM for using AWS IAM permissions, and CUSTOM for using a Lambda authorizer.

apiKeyRequired :: Maybe Bool

Specifies whether an API key is required for this route. Supported only for WebSocket APIs.

routeKey :: Maybe Text

The route key for the route.

target :: Maybe Text

The target for the route.

httpStatus :: Int

The response's http status code.

newCreateRouteResponse'

Arguments

:: Int
$sel:httpStatus:CreateRouteResponse''

-> CreateRouteResponse'

Create a value of CreateRouteResponse' with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:authorizationScopes:CreateRoute', createRouteResponse'_authorizationScopes - A list of authorization scopes configured on a route. The scopes are used with a JWT authorizer to authorize the method invocation. The authorization works by matching the route scopes against the scopes parsed from the access token in the incoming request. The method invocation is authorized if any route scope matches a claimed scope in the access token. Otherwise, the invocation is not authorized. When the route scope is configured, the client must provide an access token instead of an identity token for authorization purposes.

$sel:modelSelectionExpression:CreateRoute', createRouteResponse'_modelSelectionExpression - The model selection expression for the route. Supported only for WebSocket APIs.

$sel:requestModels:CreateRoute', createRouteResponse'_requestModels - The request models for the route. Supported only for WebSocket APIs.

$sel:routeResponseSelectionExpression:CreateRoute', createRouteResponse'_routeResponseSelectionExpression - The route response selection expression for the route. Supported only for WebSocket APIs.

$sel:requestParameters:CreateRoute', createRouteResponse'_requestParameters - The request parameters for the route. Supported only for WebSocket APIs.

$sel:routeId:CreateRouteResponse'', createRouteResponse'_routeId - The route ID.

$sel:authorizerId:CreateRoute', createRouteResponse'_authorizerId - The identifier of the Authorizer resource to be associated with this route. The authorizer identifier is generated by API Gateway when you created the authorizer.

$sel:operationName:CreateRoute', createRouteResponse'_operationName - The operation name for the route.

$sel:apiGatewayManaged:CreateRouteResponse'', createRouteResponse'_apiGatewayManaged - Specifies whether a route is managed by API Gateway. If you created an API using quick create, the $default route is managed by API Gateway. You can't modify the $default route key.

$sel:authorizationType:CreateRoute', createRouteResponse'_authorizationType - The authorization type for the route. For WebSocket APIs, valid values are NONE for open access, AWS_IAM for using AWS IAM permissions, and CUSTOM for using a Lambda authorizer For HTTP APIs, valid values are NONE for open access, JWT for using JSON Web Tokens, AWS_IAM for using AWS IAM permissions, and CUSTOM for using a Lambda authorizer.

$sel:apiKeyRequired:CreateRoute', createRouteResponse'_apiKeyRequired - Specifies whether an API key is required for this route. Supported only for WebSocket APIs.

$sel:routeKey:CreateRoute', createRouteResponse'_routeKey - The route key for the route.

$sel:target:CreateRoute', createRouteResponse'_target - The target for the route.

$sel:httpStatus:CreateRouteResponse'', createRouteResponse'_httpStatus - The response's http status code.

Response Lenses

createRouteResponse'_authorizationScopes :: Lens' CreateRouteResponse' (Maybe [Text])

A list of authorization scopes configured on a route. The scopes are used with a JWT authorizer to authorize the method invocation. The authorization works by matching the route scopes against the scopes parsed from the access token in the incoming request. The method invocation is authorized if any route scope matches a claimed scope in the access token. Otherwise, the invocation is not authorized. When the route scope is configured, the client must provide an access token instead of an identity token for authorization purposes.

createRouteResponse'_modelSelectionExpression :: Lens' CreateRouteResponse' (Maybe Text)

The model selection expression for the route. Supported only for WebSocket APIs.

createRouteResponse'_requestModels :: Lens' CreateRouteResponse' (Maybe (HashMap Text Text))

The request models for the route. Supported only for WebSocket APIs.

createRouteResponse'_routeResponseSelectionExpression :: Lens' CreateRouteResponse' (Maybe Text)

The route response selection expression for the route. Supported only for WebSocket APIs.

createRouteResponse'_requestParameters :: Lens' CreateRouteResponse' (Maybe (HashMap Text ParameterConstraints))

The request parameters for the route. Supported only for WebSocket APIs.

createRouteResponse'_routeId :: Lens' CreateRouteResponse' (Maybe Text)

The route ID.

createRouteResponse'_authorizerId :: Lens' CreateRouteResponse' (Maybe Text)

The identifier of the Authorizer resource to be associated with this route. The authorizer identifier is generated by API Gateway when you created the authorizer.

createRouteResponse'_operationName :: Lens' CreateRouteResponse' (Maybe Text)

The operation name for the route.

createRouteResponse'_apiGatewayManaged :: Lens' CreateRouteResponse' (Maybe Bool)

Specifies whether a route is managed by API Gateway. If you created an API using quick create, the $default route is managed by API Gateway. You can't modify the $default route key.

createRouteResponse'_authorizationType :: Lens' CreateRouteResponse' (Maybe AuthorizationType)

The authorization type for the route. For WebSocket APIs, valid values are NONE for open access, AWS_IAM for using AWS IAM permissions, and CUSTOM for using a Lambda authorizer For HTTP APIs, valid values are NONE for open access, JWT for using JSON Web Tokens, AWS_IAM for using AWS IAM permissions, and CUSTOM for using a Lambda authorizer.

createRouteResponse'_apiKeyRequired :: Lens' CreateRouteResponse' (Maybe Bool)

Specifies whether an API key is required for this route. Supported only for WebSocket APIs.

createRouteResponse'_routeKey :: Lens' CreateRouteResponse' (Maybe Text)

The route key for the route.

createRouteResponse'_target :: Lens' CreateRouteResponse' (Maybe Text)

The target for the route.

createRouteResponse'_httpStatus :: Lens' CreateRouteResponse' Int

The response's http status code.