Amazonka.ApiGatewayV2.CreateStage

Description

Creates a Stage for an API.

Creating a Request

data CreateStage

Creates a new Stage resource to represent a stage.

See: newCreateStage smart constructor.

Constructors

CreateStage'

Fields

deploymentId :: Maybe Text

The deployment identifier of the API stage.

routeSettings :: Maybe (HashMap Text RouteSettings)

Route settings for the stage, by routeKey.

accessLogSettings :: Maybe AccessLogSettings

Settings for logging access in this stage.

clientCertificateId :: Maybe Text

The identifier of a client certificate for a Stage. Supported only for WebSocket APIs.

stageVariables :: Maybe (HashMap Text Text)

A map that defines the stage variables for a Stage. Variable names can have alphanumeric and underscore characters, and the values must match [A-Za-z0-9-._~:/?#&=,]+.

autoDeploy :: Maybe Bool

Specifies whether updates to an API automatically trigger a new deployment. The default value is false.

defaultRouteSettings :: Maybe RouteSettings

The default route settings for the stage.

description :: Maybe Text

The description for the API stage.

tags :: Maybe (HashMap Text Text)

The collection of tags. Each tag element is associated with a given resource.

apiId :: Text

The API identifier.

stageName :: Text

The name of the stage.

newCreateStage

Arguments

:: Text
$sel:apiId:CreateStage'

-> Text
$sel:stageName:CreateStage'

-> CreateStage

Create a value of CreateStage with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:deploymentId:CreateStage', createStage_deploymentId - The deployment identifier of the API stage.

$sel:routeSettings:CreateStage', createStage_routeSettings - Route settings for the stage, by routeKey.

$sel:accessLogSettings:CreateStage', createStage_accessLogSettings - Settings for logging access in this stage.

$sel:clientCertificateId:CreateStage', createStage_clientCertificateId - The identifier of a client certificate for a Stage. Supported only for WebSocket APIs.

$sel:stageVariables:CreateStage', createStage_stageVariables - A map that defines the stage variables for a Stage. Variable names can have alphanumeric and underscore characters, and the values must match [A-Za-z0-9-._~:/?#&=,]+.

$sel:autoDeploy:CreateStage', createStage_autoDeploy - Specifies whether updates to an API automatically trigger a new deployment. The default value is false.

$sel:defaultRouteSettings:CreateStage', createStage_defaultRouteSettings - The default route settings for the stage.

$sel:description:CreateStage', createStage_description - The description for the API stage.

$sel:tags:CreateStage', createStage_tags - The collection of tags. Each tag element is associated with a given resource.

$sel:apiId:CreateStage', createStage_apiId - The API identifier.

$sel:stageName:CreateStage', createStage_stageName - The name of the stage.

Request Lenses

createStage_deploymentId :: Lens' CreateStage (Maybe Text)

The deployment identifier of the API stage.

createStage_routeSettings :: Lens' CreateStage (Maybe (HashMap Text RouteSettings))

Route settings for the stage, by routeKey.

createStage_accessLogSettings :: Lens' CreateStage (Maybe AccessLogSettings)

Settings for logging access in this stage.

createStage_clientCertificateId :: Lens' CreateStage (Maybe Text)

The identifier of a client certificate for a Stage. Supported only for WebSocket APIs.

createStage_stageVariables :: Lens' CreateStage (Maybe (HashMap Text Text))

A map that defines the stage variables for a Stage. Variable names can have alphanumeric and underscore characters, and the values must match [A-Za-z0-9-._~:/?#&=,]+.

createStage_autoDeploy :: Lens' CreateStage (Maybe Bool)

Specifies whether updates to an API automatically trigger a new deployment. The default value is false.

createStage_defaultRouteSettings :: Lens' CreateStage (Maybe RouteSettings)

The default route settings for the stage.

createStage_description :: Lens' CreateStage (Maybe Text)

The description for the API stage.

createStage_tags :: Lens' CreateStage (Maybe (HashMap Text Text))

The collection of tags. Each tag element is associated with a given resource.

createStage_apiId :: Lens' CreateStage Text

The API identifier.

createStage_stageName :: Lens' CreateStage Text

The name of the stage.

Destructuring the Response

data CreateStageResponse

See: newCreateStageResponse smart constructor.

Constructors

CreateStageResponse'

Fields

lastDeploymentStatusMessage :: Maybe Text

Describes the status of the last deployment of a stage. Supported only for stages with autoDeploy enabled.

deploymentId :: Maybe Text

The identifier of the Deployment that the Stage is associated with. Can't be updated if autoDeploy is enabled.

routeSettings :: Maybe (HashMap Text RouteSettings)

Route settings for the stage, by routeKey.

accessLogSettings :: Maybe AccessLogSettings

Settings for logging access in this stage.

clientCertificateId :: Maybe Text

The identifier of a client certificate for a Stage. Supported only for WebSocket APIs.

stageVariables :: Maybe (HashMap Text Text)

A map that defines the stage variables for a stage resource. Variable names can have alphanumeric and underscore characters, and the values must match [A-Za-z0-9-._~:/?#&=,]+.

autoDeploy :: Maybe Bool

Specifies whether updates to an API automatically trigger a new deployment. The default value is false.

createdDate :: Maybe POSIX

The timestamp when the stage was created.

defaultRouteSettings :: Maybe RouteSettings

Default route settings for the stage.

apiGatewayManaged :: Maybe Bool

Specifies whether a stage is managed by API Gateway. If you created an API using quick create, the $default stage is managed by API Gateway. You can't modify the $default stage.

stageName :: Maybe Text

The name of the stage.

lastUpdatedDate :: Maybe POSIX

The timestamp when the stage was last updated.

description :: Maybe Text

The description of the stage.

tags :: Maybe (HashMap Text Text)

The collection of tags. Each tag element is associated with a given resource.

httpStatus :: Int

The response's http status code.

newCreateStageResponse

Arguments

:: Int
$sel:httpStatus:CreateStageResponse'

-> CreateStageResponse

Create a value of CreateStageResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:lastDeploymentStatusMessage:CreateStageResponse', createStageResponse_lastDeploymentStatusMessage - Describes the status of the last deployment of a stage. Supported only for stages with autoDeploy enabled.

$sel:deploymentId:CreateStage', createStageResponse_deploymentId - The identifier of the Deployment that the Stage is associated with. Can't be updated if autoDeploy is enabled.

$sel:routeSettings:CreateStage', createStageResponse_routeSettings - Route settings for the stage, by routeKey.

$sel:accessLogSettings:CreateStage', createStageResponse_accessLogSettings - Settings for logging access in this stage.

$sel:clientCertificateId:CreateStage', createStageResponse_clientCertificateId - The identifier of a client certificate for a Stage. Supported only for WebSocket APIs.

$sel:stageVariables:CreateStage', createStageResponse_stageVariables - A map that defines the stage variables for a stage resource. Variable names can have alphanumeric and underscore characters, and the values must match [A-Za-z0-9-._~:/?#&=,]+.

$sel:autoDeploy:CreateStage', createStageResponse_autoDeploy - Specifies whether updates to an API automatically trigger a new deployment. The default value is false.

$sel:createdDate:CreateStageResponse', createStageResponse_createdDate - The timestamp when the stage was created.

$sel:defaultRouteSettings:CreateStage', createStageResponse_defaultRouteSettings - Default route settings for the stage.

$sel:apiGatewayManaged:CreateStageResponse', createStageResponse_apiGatewayManaged - Specifies whether a stage is managed by API Gateway. If you created an API using quick create, the $default stage is managed by API Gateway. You can't modify the $default stage.

$sel:stageName:CreateStage', createStageResponse_stageName - The name of the stage.

$sel:lastUpdatedDate:CreateStageResponse', createStageResponse_lastUpdatedDate - The timestamp when the stage was last updated.

$sel:description:CreateStage', createStageResponse_description - The description of the stage.

$sel:tags:CreateStage', createStageResponse_tags - The collection of tags. Each tag element is associated with a given resource.

$sel:httpStatus:CreateStageResponse', createStageResponse_httpStatus - The response's http status code.

Response Lenses

createStageResponse_lastDeploymentStatusMessage :: Lens' CreateStageResponse (Maybe Text)

Describes the status of the last deployment of a stage. Supported only for stages with autoDeploy enabled.

createStageResponse_deploymentId :: Lens' CreateStageResponse (Maybe Text)

The identifier of the Deployment that the Stage is associated with. Can't be updated if autoDeploy is enabled.

createStageResponse_routeSettings :: Lens' CreateStageResponse (Maybe (HashMap Text RouteSettings))

Route settings for the stage, by routeKey.

createStageResponse_accessLogSettings :: Lens' CreateStageResponse (Maybe AccessLogSettings)

Settings for logging access in this stage.

createStageResponse_clientCertificateId :: Lens' CreateStageResponse (Maybe Text)

The identifier of a client certificate for a Stage. Supported only for WebSocket APIs.

createStageResponse_stageVariables :: Lens' CreateStageResponse (Maybe (HashMap Text Text))

A map that defines the stage variables for a stage resource. Variable names can have alphanumeric and underscore characters, and the values must match [A-Za-z0-9-._~:/?#&=,]+.

createStageResponse_autoDeploy :: Lens' CreateStageResponse (Maybe Bool)

Specifies whether updates to an API automatically trigger a new deployment. The default value is false.

createStageResponse_createdDate :: Lens' CreateStageResponse (Maybe UTCTime)

The timestamp when the stage was created.

createStageResponse_defaultRouteSettings :: Lens' CreateStageResponse (Maybe RouteSettings)

Default route settings for the stage.

createStageResponse_apiGatewayManaged :: Lens' CreateStageResponse (Maybe Bool)

Specifies whether a stage is managed by API Gateway. If you created an API using quick create, the $default stage is managed by API Gateway. You can't modify the $default stage.

createStageResponse_stageName :: Lens' CreateStageResponse (Maybe Text)

The name of the stage.

createStageResponse_lastUpdatedDate :: Lens' CreateStageResponse (Maybe UTCTime)

The timestamp when the stage was last updated.

createStageResponse_description :: Lens' CreateStageResponse (Maybe Text)

The description of the stage.

createStageResponse_tags :: Lens' CreateStageResponse (Maybe (HashMap Text Text))

The collection of tags. Each tag element is associated with a given resource.

createStageResponse_httpStatus :: Lens' CreateStageResponse Int

The response's http status code.