Amazonka.ApiGatewayV2.GetDeployment

Description

Gets a Deployment.

Creating a Request

data GetDeployment

See: newGetDeployment smart constructor.

Constructors

GetDeployment'

Fields

apiId :: Text

The API identifier.

deploymentId :: Text

The deployment ID.

newGetDeployment

Arguments

:: Text
$sel:apiId:GetDeployment'

-> Text
$sel:deploymentId:GetDeployment'

-> GetDeployment

Create a value of GetDeployment with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:apiId:GetDeployment', getDeployment_apiId - The API identifier.

$sel:deploymentId:GetDeployment', getDeployment_deploymentId - The deployment ID.

Request Lenses

getDeployment_apiId :: Lens' GetDeployment Text

The API identifier.

getDeployment_deploymentId :: Lens' GetDeployment Text

The deployment ID.

Destructuring the Response

data GetDeploymentResponse

See: newGetDeploymentResponse smart constructor.

Constructors

GetDeploymentResponse'

Fields

deploymentId :: Maybe Text

The identifier for the deployment.

autoDeployed :: Maybe Bool

Specifies whether a deployment was automatically released.

deploymentStatusMessage :: Maybe Text

May contain additional feedback on the status of an API deployment.

createdDate :: Maybe POSIX

The date and time when the Deployment resource was created.

deploymentStatus :: Maybe DeploymentStatus

The status of the deployment: PENDING, FAILED, or SUCCEEDED.

description :: Maybe Text

The description for the deployment.

httpStatus :: Int

The response's http status code.

newGetDeploymentResponse

Arguments

:: Int
$sel:httpStatus:GetDeploymentResponse'

-> GetDeploymentResponse

Create a value of GetDeploymentResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:deploymentId:GetDeployment', getDeploymentResponse_deploymentId - The identifier for the deployment.

$sel:autoDeployed:GetDeploymentResponse', getDeploymentResponse_autoDeployed - Specifies whether a deployment was automatically released.

$sel:deploymentStatusMessage:GetDeploymentResponse', getDeploymentResponse_deploymentStatusMessage - May contain additional feedback on the status of an API deployment.

$sel:createdDate:GetDeploymentResponse', getDeploymentResponse_createdDate - The date and time when the Deployment resource was created.

$sel:deploymentStatus:GetDeploymentResponse', getDeploymentResponse_deploymentStatus - The status of the deployment: PENDING, FAILED, or SUCCEEDED.

$sel:description:GetDeploymentResponse', getDeploymentResponse_description - The description for the deployment.

$sel:httpStatus:GetDeploymentResponse', getDeploymentResponse_httpStatus - The response's http status code.

Response Lenses

getDeploymentResponse_deploymentId :: Lens' GetDeploymentResponse (Maybe Text)

The identifier for the deployment.

getDeploymentResponse_autoDeployed :: Lens' GetDeploymentResponse (Maybe Bool)

Specifies whether a deployment was automatically released.

getDeploymentResponse_deploymentStatusMessage :: Lens' GetDeploymentResponse (Maybe Text)

May contain additional feedback on the status of an API deployment.

getDeploymentResponse_createdDate :: Lens' GetDeploymentResponse (Maybe UTCTime)

The date and time when the Deployment resource was created.

getDeploymentResponse_deploymentStatus :: Lens' GetDeploymentResponse (Maybe DeploymentStatus)

The status of the deployment: PENDING, FAILED, or SUCCEEDED.

getDeploymentResponse_description :: Lens' GetDeploymentResponse (Maybe Text)

The description for the deployment.

getDeploymentResponse_httpStatus :: Lens' GetDeploymentResponse Int

The response's http status code.