Amazonka.ApiGatewayV2.GetDeployments

Description

Gets the Deployments for an API.

This operation returns paginated results.

Creating a Request

data GetDeployments

See: newGetDeployments smart constructor.

Constructors

GetDeployments'

Fields

nextToken :: Maybe Text

The next page of elements from this collection. Not valid for the last element of the collection.

maxResults :: Maybe Text

The maximum number of elements to be returned for this resource.

apiId :: Text

The API identifier.

newGetDeployments

Arguments

:: Text
$sel:apiId:GetDeployments'

-> GetDeployments

Create a value of GetDeployments with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:nextToken:GetDeployments', getDeployments_nextToken - The next page of elements from this collection. Not valid for the last element of the collection.

$sel:maxResults:GetDeployments', getDeployments_maxResults - The maximum number of elements to be returned for this resource.

$sel:apiId:GetDeployments', getDeployments_apiId - The API identifier.

Request Lenses

getDeployments_nextToken :: Lens' GetDeployments (Maybe Text)

The next page of elements from this collection. Not valid for the last element of the collection.

getDeployments_maxResults :: Lens' GetDeployments (Maybe Text)

The maximum number of elements to be returned for this resource.

getDeployments_apiId :: Lens' GetDeployments Text

The API identifier.

Destructuring the Response

data GetDeploymentsResponse

See: newGetDeploymentsResponse smart constructor.

Constructors

GetDeploymentsResponse'

Fields

items :: Maybe [Deployment]

The elements from this collection.

nextToken :: Maybe Text

The next page of elements from this collection. Not valid for the last element of the collection.

httpStatus :: Int

The response's http status code.

newGetDeploymentsResponse

Arguments

:: Int
$sel:httpStatus:GetDeploymentsResponse'

-> GetDeploymentsResponse

Create a value of GetDeploymentsResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:items:GetDeploymentsResponse', getDeploymentsResponse_items - The elements from this collection.

$sel:nextToken:GetDeployments', getDeploymentsResponse_nextToken - The next page of elements from this collection. Not valid for the last element of the collection.

$sel:httpStatus:GetDeploymentsResponse', getDeploymentsResponse_httpStatus - The response's http status code.

Response Lenses

getDeploymentsResponse_items :: Lens' GetDeploymentsResponse (Maybe [Deployment])

The elements from this collection.

getDeploymentsResponse_nextToken :: Lens' GetDeploymentsResponse (Maybe Text)

The next page of elements from this collection. Not valid for the last element of the collection.

getDeploymentsResponse_httpStatus :: Lens' GetDeploymentsResponse Int

The response's http status code.