Amazonka.ApiGatewayV2.GetDomainName

Description

Gets a domain name.

Creating a Request

data GetDomainName

See: newGetDomainName smart constructor.

Constructors

GetDomainName'

Fields

domainName :: Text

The domain name.

newGetDomainName

Arguments

:: Text
$sel:domainName:GetDomainName'

-> GetDomainName

Create a value of GetDomainName with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:domainName:GetDomainName', getDomainName_domainName - The domain name.

Request Lenses

getDomainName_domainName :: Lens' GetDomainName Text

The domain name.

Destructuring the Response

data GetDomainNameResponse

See: newGetDomainNameResponse smart constructor.

Constructors

GetDomainNameResponse'

Fields

domainNameConfigurations :: Maybe [DomainNameConfiguration]

The domain name configurations.

domainName :: Maybe Text

The name of the DomainName resource.

mutualTlsAuthentication :: Maybe MutualTlsAuthentication

The mutual TLS authentication configuration for a custom domain name.

apiMappingSelectionExpression :: Maybe Text

The API mapping selection expression.

tags :: Maybe (HashMap Text Text)

The collection of tags associated with a domain name.

httpStatus :: Int

The response's http status code.

newGetDomainNameResponse

Arguments

:: Int
$sel:httpStatus:GetDomainNameResponse'

-> GetDomainNameResponse

Create a value of GetDomainNameResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:domainNameConfigurations:GetDomainNameResponse', getDomainNameResponse_domainNameConfigurations - The domain name configurations.

$sel:domainName:GetDomainName', getDomainNameResponse_domainName - The name of the DomainName resource.

$sel:mutualTlsAuthentication:GetDomainNameResponse', getDomainNameResponse_mutualTlsAuthentication - The mutual TLS authentication configuration for a custom domain name.

$sel:apiMappingSelectionExpression:GetDomainNameResponse', getDomainNameResponse_apiMappingSelectionExpression - The API mapping selection expression.

$sel:tags:GetDomainNameResponse', getDomainNameResponse_tags - The collection of tags associated with a domain name.

$sel:httpStatus:GetDomainNameResponse', getDomainNameResponse_httpStatus - The response's http status code.

Response Lenses

getDomainNameResponse_domainNameConfigurations :: Lens' GetDomainNameResponse (Maybe [DomainNameConfiguration])

The domain name configurations.

getDomainNameResponse_domainName :: Lens' GetDomainNameResponse (Maybe Text)

The name of the DomainName resource.

getDomainNameResponse_mutualTlsAuthentication :: Lens' GetDomainNameResponse (Maybe MutualTlsAuthentication)

The mutual TLS authentication configuration for a custom domain name.

getDomainNameResponse_apiMappingSelectionExpression :: Lens' GetDomainNameResponse (Maybe Text)

The API mapping selection expression.

getDomainNameResponse_tags :: Lens' GetDomainNameResponse (Maybe (HashMap Text Text))

The collection of tags associated with a domain name.

getDomainNameResponse_httpStatus :: Lens' GetDomainNameResponse Int

The response's http status code.