Amazonka.ApiGatewayV2.GetDomainNames

Description

Gets the domain names for an AWS account.

This operation returns paginated results.

Creating a Request

data GetDomainNames

See: newGetDomainNames smart constructor.

Constructors

GetDomainNames'

Fields

nextToken :: Maybe Text

The next page of elements from this collection. Not valid for the last element of the collection.

maxResults :: Maybe Text

The maximum number of elements to be returned for this resource.

newGetDomainNames :: GetDomainNames

Create a value of GetDomainNames with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:nextToken:GetDomainNames', getDomainNames_nextToken - The next page of elements from this collection. Not valid for the last element of the collection.

$sel:maxResults:GetDomainNames', getDomainNames_maxResults - The maximum number of elements to be returned for this resource.

Request Lenses

getDomainNames_nextToken :: Lens' GetDomainNames (Maybe Text)

The next page of elements from this collection. Not valid for the last element of the collection.

getDomainNames_maxResults :: Lens' GetDomainNames (Maybe Text)

The maximum number of elements to be returned for this resource.

Destructuring the Response

data GetDomainNamesResponse

See: newGetDomainNamesResponse smart constructor.

Constructors

GetDomainNamesResponse'

Fields

items :: Maybe [DomainName]

The elements from this collection.

nextToken :: Maybe Text

The next page of elements from this collection. Not valid for the last element of the collection.

httpStatus :: Int

The response's http status code.

newGetDomainNamesResponse

Arguments

:: Int
$sel:httpStatus:GetDomainNamesResponse'

-> GetDomainNamesResponse

Create a value of GetDomainNamesResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:items:GetDomainNamesResponse', getDomainNamesResponse_items - The elements from this collection.

$sel:nextToken:GetDomainNames', getDomainNamesResponse_nextToken - The next page of elements from this collection. Not valid for the last element of the collection.

$sel:httpStatus:GetDomainNamesResponse', getDomainNamesResponse_httpStatus - The response's http status code.

Response Lenses

getDomainNamesResponse_items :: Lens' GetDomainNamesResponse (Maybe [DomainName])

The elements from this collection.

getDomainNamesResponse_nextToken :: Lens' GetDomainNamesResponse (Maybe Text)

The next page of elements from this collection. Not valid for the last element of the collection.

getDomainNamesResponse_httpStatus :: Lens' GetDomainNamesResponse Int

The response's http status code.