Amazonka.ApiGatewayV2.GetIntegrationResponse

Description

Gets an IntegrationResponses.

Creating a Request

data GetIntegrationResponse

See: newGetIntegrationResponse smart constructor.

Constructors

GetIntegrationResponse'

Fields

apiId :: Text

The API identifier.

integrationResponseId :: Text

The integration response ID.

integrationId :: Text

The integration ID.

newGetIntegrationResponse

Arguments

:: Text
$sel:apiId:GetIntegrationResponse'

-> Text
$sel:integrationResponseId:GetIntegrationResponse'

-> Text
$sel:integrationId:GetIntegrationResponse'

-> GetIntegrationResponse

Create a value of GetIntegrationResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:apiId:GetIntegrationResponse', getIntegrationResponse_apiId - The API identifier.

$sel:integrationResponseId:GetIntegrationResponse', getIntegrationResponse_integrationResponseId - The integration response ID.

$sel:integrationId:GetIntegrationResponse', getIntegrationResponse_integrationId - The integration ID.

Request Lenses

getIntegrationResponse_apiId :: Lens' GetIntegrationResponse Text

The API identifier.

getIntegrationResponse_integrationResponseId :: Lens' GetIntegrationResponse Text

The integration response ID.

getIntegrationResponse_integrationId :: Lens' GetIntegrationResponse Text

The integration ID.

Destructuring the Response

data GetIntegrationResponseResponse

See: newGetIntegrationResponseResponse smart constructor.

Constructors

GetIntegrationResponseResponse'

Fields

integrationResponseId :: Maybe Text

The integration response ID.

integrationResponseKey :: Maybe Text

The integration response key.

templateSelectionExpression :: Maybe Text

The template selection expressions for the integration response.

contentHandlingStrategy :: Maybe ContentHandlingStrategy

Supported only for WebSocket APIs. Specifies how to handle response payload content type conversions. Supported values are CONVERT_TO_BINARY and CONVERT_TO_TEXT, with the following behaviors:

CONVERT_TO_BINARY: Converts a response payload from a Base64-encoded string to the corresponding binary blob.

CONVERT_TO_TEXT: Converts a response payload from a binary blob to a Base64-encoded string.

If this property is not defined, the response payload will be passed through from the integration response to the route response or method response without modification.

responseTemplates :: Maybe (HashMap Text Text)

The collection of response templates for the integration response as a string-to-string map of key-value pairs. Response templates are represented as a key/value map, with a content-type as the key and a template as the value.

responseParameters :: Maybe (HashMap Text Text)

A key-value map specifying response parameters that are passed to the method response from the backend. The key is a method response header parameter name and the mapped value is an integration response header value, a static value enclosed within a pair of single quotes, or a JSON expression from the integration response body. The mapping key must match the pattern of method.response.header.{name}, where name is a valid and unique header name. The mapped non-static value must match the pattern of integration.response.header.{name} or integration.response.body.{JSON-expression}, where name is a valid and unique response header name and JSON-expression is a valid JSON expression without the $ prefix.

httpStatus :: Int

The response's http status code.

newGetIntegrationResponseResponse

Arguments

:: Int
$sel:httpStatus:GetIntegrationResponseResponse'

-> GetIntegrationResponseResponse

Create a value of GetIntegrationResponseResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:integrationResponseId:GetIntegrationResponse', getIntegrationResponseResponse_integrationResponseId - The integration response ID.

$sel:integrationResponseKey:GetIntegrationResponseResponse', getIntegrationResponseResponse_integrationResponseKey - The integration response key.

$sel:templateSelectionExpression:GetIntegrationResponseResponse', getIntegrationResponseResponse_templateSelectionExpression - The template selection expressions for the integration response.

$sel:contentHandlingStrategy:GetIntegrationResponseResponse', getIntegrationResponseResponse_contentHandlingStrategy - Supported only for WebSocket APIs. Specifies how to handle response payload content type conversions. Supported values are CONVERT_TO_BINARY and CONVERT_TO_TEXT, with the following behaviors:

CONVERT_TO_BINARY: Converts a response payload from a Base64-encoded string to the corresponding binary blob.

CONVERT_TO_TEXT: Converts a response payload from a binary blob to a Base64-encoded string.

If this property is not defined, the response payload will be passed through from the integration response to the route response or method response without modification.

$sel:responseTemplates:GetIntegrationResponseResponse', getIntegrationResponseResponse_responseTemplates - The collection of response templates for the integration response as a string-to-string map of key-value pairs. Response templates are represented as a key/value map, with a content-type as the key and a template as the value.

$sel:responseParameters:GetIntegrationResponseResponse', getIntegrationResponseResponse_responseParameters - A key-value map specifying response parameters that are passed to the method response from the backend. The key is a method response header parameter name and the mapped value is an integration response header value, a static value enclosed within a pair of single quotes, or a JSON expression from the integration response body. The mapping key must match the pattern of method.response.header.{name}, where name is a valid and unique header name. The mapped non-static value must match the pattern of integration.response.header.{name} or integration.response.body.{JSON-expression}, where name is a valid and unique response header name and JSON-expression is a valid JSON expression without the $ prefix.

$sel:httpStatus:GetIntegrationResponseResponse', getIntegrationResponseResponse_httpStatus - The response's http status code.

Response Lenses

getIntegrationResponseResponse_integrationResponseId :: Lens' GetIntegrationResponseResponse (Maybe Text)

The integration response ID.

getIntegrationResponseResponse_integrationResponseKey :: Lens' GetIntegrationResponseResponse (Maybe Text)

The integration response key.

getIntegrationResponseResponse_templateSelectionExpression :: Lens' GetIntegrationResponseResponse (Maybe Text)

The template selection expressions for the integration response.

getIntegrationResponseResponse_contentHandlingStrategy :: Lens' GetIntegrationResponseResponse (Maybe ContentHandlingStrategy)

Supported only for WebSocket APIs. Specifies how to handle response payload content type conversions. Supported values are CONVERT_TO_BINARY and CONVERT_TO_TEXT, with the following behaviors:

CONVERT_TO_BINARY: Converts a response payload from a Base64-encoded string to the corresponding binary blob.

CONVERT_TO_TEXT: Converts a response payload from a binary blob to a Base64-encoded string.

If this property is not defined, the response payload will be passed through from the integration response to the route response or method response without modification.

getIntegrationResponseResponse_responseTemplates :: Lens' GetIntegrationResponseResponse (Maybe (HashMap Text Text))

The collection of response templates for the integration response as a string-to-string map of key-value pairs. Response templates are represented as a key/value map, with a content-type as the key and a template as the value.

getIntegrationResponseResponse_responseParameters :: Lens' GetIntegrationResponseResponse (Maybe (HashMap Text Text))

A key-value map specifying response parameters that are passed to the method response from the backend. The key is a method response header parameter name and the mapped value is an integration response header value, a static value enclosed within a pair of single quotes, or a JSON expression from the integration response body. The mapping key must match the pattern of method.response.header.{name}, where name is a valid and unique header name. The mapped non-static value must match the pattern of integration.response.header.{name} or integration.response.body.{JSON-expression}, where name is a valid and unique response header name and JSON-expression is a valid JSON expression without the $ prefix.

getIntegrationResponseResponse_httpStatus :: Lens' GetIntegrationResponseResponse Int

The response's http status code.