Amazonka.ApiGatewayV2.GetModel

Description

Gets a Model.

Creating a Request

data GetModel

See: newGetModel smart constructor.

Constructors

GetModel'

Fields

modelId :: Text

The model ID.

apiId :: Text

The API identifier.

newGetModel

Arguments

:: Text
$sel:modelId:GetModel'

-> Text
$sel:apiId:GetModel'

-> GetModel

Create a value of GetModel with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:modelId:GetModel', getModel_modelId - The model ID.

$sel:apiId:GetModel', getModel_apiId - The API identifier.

Request Lenses

getModel_modelId :: Lens' GetModel Text

The model ID.

getModel_apiId :: Lens' GetModel Text

The API identifier.

Destructuring the Response

data GetModelResponse

See: newGetModelResponse smart constructor.

Constructors

GetModelResponse'

Fields

modelId :: Maybe Text

The model identifier.

schema :: Maybe Text

The schema for the model. For application/json models, this should be JSON schema draft 4 model.

name :: Maybe Text

The name of the model. Must be alphanumeric.

description :: Maybe Text

The description of the model.

contentType :: Maybe Text

The content-type for the model, for example, "application/json".

httpStatus :: Int

The response's http status code.

newGetModelResponse

Arguments

:: Int
$sel:httpStatus:GetModelResponse'

-> GetModelResponse

Create a value of GetModelResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:modelId:GetModel', getModelResponse_modelId - The model identifier.

$sel:schema:GetModelResponse', getModelResponse_schema - The schema for the model. For application/json models, this should be JSON schema draft 4 model.

$sel:name:GetModelResponse', getModelResponse_name - The name of the model. Must be alphanumeric.

$sel:description:GetModelResponse', getModelResponse_description - The description of the model.

$sel:contentType:GetModelResponse', getModelResponse_contentType - The content-type for the model, for example, "application/json".

$sel:httpStatus:GetModelResponse', getModelResponse_httpStatus - The response's http status code.

Response Lenses

getModelResponse_modelId :: Lens' GetModelResponse (Maybe Text)

The model identifier.

getModelResponse_schema :: Lens' GetModelResponse (Maybe Text)

The schema for the model. For application/json models, this should be JSON schema draft 4 model.

getModelResponse_name :: Lens' GetModelResponse (Maybe Text)

The name of the model. Must be alphanumeric.

getModelResponse_description :: Lens' GetModelResponse (Maybe Text)

The description of the model.

getModelResponse_contentType :: Lens' GetModelResponse (Maybe Text)

The content-type for the model, for example, "application/json".

getModelResponse_httpStatus :: Lens' GetModelResponse Int

The response's http status code.