Amazonka.ApiGatewayV2.GetStages

Description

Gets the Stages for an API.

This operation returns paginated results.

Creating a Request

data GetStages

See: newGetStages smart constructor.

Constructors

GetStages'

Fields

nextToken :: Maybe Text

The next page of elements from this collection. Not valid for the last element of the collection.

maxResults :: Maybe Text

The maximum number of elements to be returned for this resource.

apiId :: Text

The API identifier.

newGetStages

Arguments

:: Text
$sel:apiId:GetStages'

-> GetStages

Create a value of GetStages with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:nextToken:GetStages', getStages_nextToken - The next page of elements from this collection. Not valid for the last element of the collection.

$sel:maxResults:GetStages', getStages_maxResults - The maximum number of elements to be returned for this resource.

$sel:apiId:GetStages', getStages_apiId - The API identifier.

Request Lenses

getStages_nextToken :: Lens' GetStages (Maybe Text)

The next page of elements from this collection. Not valid for the last element of the collection.

getStages_maxResults :: Lens' GetStages (Maybe Text)

The maximum number of elements to be returned for this resource.

getStages_apiId :: Lens' GetStages Text

The API identifier.

Destructuring the Response

data GetStagesResponse

See: newGetStagesResponse smart constructor.

Constructors

GetStagesResponse'

Fields

items :: Maybe [Stage]

The elements from this collection.

nextToken :: Maybe Text

The next page of elements from this collection. Not valid for the last element of the collection.

httpStatus :: Int

The response's http status code.

newGetStagesResponse

Arguments

:: Int
$sel:httpStatus:GetStagesResponse'

-> GetStagesResponse

Create a value of GetStagesResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:items:GetStagesResponse', getStagesResponse_items - The elements from this collection.

$sel:nextToken:GetStages', getStagesResponse_nextToken - The next page of elements from this collection. Not valid for the last element of the collection.

$sel:httpStatus:GetStagesResponse', getStagesResponse_httpStatus - The response's http status code.

Response Lenses

getStagesResponse_items :: Lens' GetStagesResponse (Maybe [Stage])

The elements from this collection.

getStagesResponse_nextToken :: Lens' GetStagesResponse (Maybe Text)

The next page of elements from this collection. Not valid for the last element of the collection.

getStagesResponse_httpStatus :: Lens' GetStagesResponse Int

The response's http status code.