Amazonka.ApiGatewayV2.ReimportApi

Description

Puts an Api resource.

Creating a Request

data ReimportApi

See: newReimportApi smart constructor.

Constructors

ReimportApi'

Fields

basepath :: Maybe Text

Specifies how to interpret the base path of the API during import. Valid values are ignore, prepend, and split. The default value is ignore. To learn more, see Set the OpenAPI basePath Property. Supported only for HTTP APIs.

failOnWarnings :: Maybe Bool

Specifies whether to rollback the API creation when a warning is encountered. By default, API creation continues if a warning is encountered.

apiId :: Text

The API identifier.

body :: Text

The OpenAPI definition. Supported only for HTTP APIs.

newReimportApi

Arguments

:: Text
$sel:apiId:ReimportApi'

-> Text
$sel:body:ReimportApi'

-> ReimportApi

Create a value of ReimportApi with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:basepath:ReimportApi', reimportApi_basepath - Specifies how to interpret the base path of the API during import. Valid values are ignore, prepend, and split. The default value is ignore. To learn more, see Set the OpenAPI basePath Property. Supported only for HTTP APIs.

$sel:failOnWarnings:ReimportApi', reimportApi_failOnWarnings - Specifies whether to rollback the API creation when a warning is encountered. By default, API creation continues if a warning is encountered.

$sel:apiId:ReimportApi', reimportApi_apiId - The API identifier.

$sel:body:ReimportApi', reimportApi_body - The OpenAPI definition. Supported only for HTTP APIs.

Request Lenses

reimportApi_basepath :: Lens' ReimportApi (Maybe Text)

Specifies how to interpret the base path of the API during import. Valid values are ignore, prepend, and split. The default value is ignore. To learn more, see Set the OpenAPI basePath Property. Supported only for HTTP APIs.

reimportApi_failOnWarnings :: Lens' ReimportApi (Maybe Bool)

Specifies whether to rollback the API creation when a warning is encountered. By default, API creation continues if a warning is encountered.

reimportApi_apiId :: Lens' ReimportApi Text

The API identifier.

reimportApi_body :: Lens' ReimportApi Text

The OpenAPI definition. Supported only for HTTP APIs.

Destructuring the Response

data ReimportApiResponse

See: newReimportApiResponse smart constructor.

Constructors

ReimportApiResponse'

Fields

apiId :: Maybe Text

The API ID.

disableExecuteApiEndpoint :: Maybe Bool

Specifies whether clients can invoke your API by using the default execute-api endpoint. By default, clients can invoke your API with the default https://{api_id}.execute-api.{region}.amazonaws.com endpoint. To require that clients use a custom domain name to invoke your API, disable the default endpoint.

apiEndpoint :: Maybe Text

The URI of the API, of the form {api-id}.execute-api.{region}.amazonaws.com. The stage name is typically appended to this URI to form a complete path to a deployed API stage.

warnings :: Maybe [Text]

The warning messages reported when failonwarnings is turned on during API import.

createdDate :: Maybe POSIX

The timestamp when the API was created.

name :: Maybe Text

The name of the API.

version :: Maybe Text

A version identifier for the API.

apiGatewayManaged :: Maybe Bool

Specifies whether an API is managed by API Gateway. You can't update or delete a managed API by using API Gateway. A managed API can be deleted only through the tooling or service that created it.

apiKeySelectionExpression :: Maybe Text

An API key selection expression. Supported only for WebSocket APIs. See API Key Selection Expressions.

corsConfiguration :: Maybe Cors

A CORS configuration. Supported only for HTTP APIs.

routeSelectionExpression :: Maybe Text

The route selection expression for the API. For HTTP APIs, the routeSelectionExpression must be ${request.method} ${request.path}. If not provided, this will be the default for HTTP APIs. This property is required for WebSocket APIs.

importInfo :: Maybe [Text]

The validation information during API import. This may include particular properties of your OpenAPI definition which are ignored during import. Supported only for HTTP APIs.

disableSchemaValidation :: Maybe Bool

Avoid validating models when creating a deployment. Supported only for WebSocket APIs.

description :: Maybe Text

The description of the API.

protocolType :: Maybe ProtocolType

The API protocol.

tags :: Maybe (HashMap Text Text)

A collection of tags associated with the API.

httpStatus :: Int

The response's http status code.

newReimportApiResponse

Arguments

:: Int
$sel:httpStatus:ReimportApiResponse'

-> ReimportApiResponse

Create a value of ReimportApiResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:apiId:ReimportApi', reimportApiResponse_apiId - The API ID.

$sel:disableExecuteApiEndpoint:ReimportApiResponse', reimportApiResponse_disableExecuteApiEndpoint - Specifies whether clients can invoke your API by using the default execute-api endpoint. By default, clients can invoke your API with the default https://{api_id}.execute-api.{region}.amazonaws.com endpoint. To require that clients use a custom domain name to invoke your API, disable the default endpoint.

$sel:apiEndpoint:ReimportApiResponse', reimportApiResponse_apiEndpoint - The URI of the API, of the form {api-id}.execute-api.{region}.amazonaws.com. The stage name is typically appended to this URI to form a complete path to a deployed API stage.

$sel:warnings:ReimportApiResponse', reimportApiResponse_warnings - The warning messages reported when failonwarnings is turned on during API import.

$sel:createdDate:ReimportApiResponse', reimportApiResponse_createdDate - The timestamp when the API was created.

$sel:name:ReimportApiResponse', reimportApiResponse_name - The name of the API.

$sel:version:ReimportApiResponse', reimportApiResponse_version - A version identifier for the API.

$sel:apiGatewayManaged:ReimportApiResponse', reimportApiResponse_apiGatewayManaged - Specifies whether an API is managed by API Gateway. You can't update or delete a managed API by using API Gateway. A managed API can be deleted only through the tooling or service that created it.

$sel:apiKeySelectionExpression:ReimportApiResponse', reimportApiResponse_apiKeySelectionExpression - An API key selection expression. Supported only for WebSocket APIs. See API Key Selection Expressions.

$sel:corsConfiguration:ReimportApiResponse', reimportApiResponse_corsConfiguration - A CORS configuration. Supported only for HTTP APIs.

$sel:routeSelectionExpression:ReimportApiResponse', reimportApiResponse_routeSelectionExpression - The route selection expression for the API. For HTTP APIs, the routeSelectionExpression must be ${request.method} ${request.path}. If not provided, this will be the default for HTTP APIs. This property is required for WebSocket APIs.

$sel:importInfo:ReimportApiResponse', reimportApiResponse_importInfo - The validation information during API import. This may include particular properties of your OpenAPI definition which are ignored during import. Supported only for HTTP APIs.

$sel:disableSchemaValidation:ReimportApiResponse', reimportApiResponse_disableSchemaValidation - Avoid validating models when creating a deployment. Supported only for WebSocket APIs.

$sel:description:ReimportApiResponse', reimportApiResponse_description - The description of the API.

$sel:protocolType:ReimportApiResponse', reimportApiResponse_protocolType - The API protocol.

$sel:tags:ReimportApiResponse', reimportApiResponse_tags - A collection of tags associated with the API.

$sel:httpStatus:ReimportApiResponse', reimportApiResponse_httpStatus - The response's http status code.

Response Lenses

reimportApiResponse_apiId :: Lens' ReimportApiResponse (Maybe Text)

The API ID.

reimportApiResponse_disableExecuteApiEndpoint :: Lens' ReimportApiResponse (Maybe Bool)

Specifies whether clients can invoke your API by using the default execute-api endpoint. By default, clients can invoke your API with the default https://{api_id}.execute-api.{region}.amazonaws.com endpoint. To require that clients use a custom domain name to invoke your API, disable the default endpoint.

reimportApiResponse_apiEndpoint :: Lens' ReimportApiResponse (Maybe Text)

The URI of the API, of the form {api-id}.execute-api.{region}.amazonaws.com. The stage name is typically appended to this URI to form a complete path to a deployed API stage.

reimportApiResponse_warnings :: Lens' ReimportApiResponse (Maybe [Text])

The warning messages reported when failonwarnings is turned on during API import.

reimportApiResponse_createdDate :: Lens' ReimportApiResponse (Maybe UTCTime)

The timestamp when the API was created.

reimportApiResponse_name :: Lens' ReimportApiResponse (Maybe Text)

The name of the API.

reimportApiResponse_version :: Lens' ReimportApiResponse (Maybe Text)

A version identifier for the API.

reimportApiResponse_apiGatewayManaged :: Lens' ReimportApiResponse (Maybe Bool)

Specifies whether an API is managed by API Gateway. You can't update or delete a managed API by using API Gateway. A managed API can be deleted only through the tooling or service that created it.

reimportApiResponse_apiKeySelectionExpression :: Lens' ReimportApiResponse (Maybe Text)

An API key selection expression. Supported only for WebSocket APIs. See API Key Selection Expressions.

reimportApiResponse_corsConfiguration :: Lens' ReimportApiResponse (Maybe Cors)

A CORS configuration. Supported only for HTTP APIs.

reimportApiResponse_routeSelectionExpression :: Lens' ReimportApiResponse (Maybe Text)

The route selection expression for the API. For HTTP APIs, the routeSelectionExpression must be ${request.method} ${request.path}. If not provided, this will be the default for HTTP APIs. This property is required for WebSocket APIs.

reimportApiResponse_importInfo :: Lens' ReimportApiResponse (Maybe [Text])

The validation information during API import. This may include particular properties of your OpenAPI definition which are ignored during import. Supported only for HTTP APIs.

reimportApiResponse_disableSchemaValidation :: Lens' ReimportApiResponse (Maybe Bool)

Avoid validating models when creating a deployment. Supported only for WebSocket APIs.

reimportApiResponse_description :: Lens' ReimportApiResponse (Maybe Text)

The description of the API.

reimportApiResponse_protocolType :: Lens' ReimportApiResponse (Maybe ProtocolType)

The API protocol.

reimportApiResponse_tags :: Lens' ReimportApiResponse (Maybe (HashMap Text Text))

A collection of tags associated with the API.

reimportApiResponse_httpStatus :: Lens' ReimportApiResponse Int

The response's http status code.