Amazonka.ApiGatewayV2.Types.Api

Description

Documentation

data Api

Represents an API.

See: newApi smart constructor.

Constructors

Api'

Fields

apiId :: Maybe Text

The API ID.

disableExecuteApiEndpoint :: Maybe Bool

Specifies whether clients can invoke your API by using the default execute-api endpoint. By default, clients can invoke your API with the default https://{api_id}.execute-api.{region}.amazonaws.com endpoint. To require that clients use a custom domain name to invoke your API, disable the default endpoint.

apiEndpoint :: Maybe Text

The URI of the API, of the form {api-id}.execute-api.{region}.amazonaws.com. The stage name is typically appended to this URI to form a complete path to a deployed API stage.

warnings :: Maybe [Text]

The warning messages reported when failonwarnings is turned on during API import.

createdDate :: Maybe POSIX

The timestamp when the API was created.

version :: Maybe Text

A version identifier for the API.

apiGatewayManaged :: Maybe Bool

Specifies whether an API is managed by API Gateway. You can't update or delete a managed API by using API Gateway. A managed API can be deleted only through the tooling or service that created it.

apiKeySelectionExpression :: Maybe Text

An API key selection expression. Supported only for WebSocket APIs. See API Key Selection Expressions.

corsConfiguration :: Maybe Cors

A CORS configuration. Supported only for HTTP APIs.

importInfo :: Maybe [Text]

The validation information during API import. This may include particular properties of your OpenAPI definition which are ignored during import. Supported only for HTTP APIs.

disableSchemaValidation :: Maybe Bool

Avoid validating models when creating a deployment. Supported only for WebSocket APIs.

description :: Maybe Text

The description of the API.

tags :: Maybe (HashMap Text Text)

A collection of tags associated with the API.

routeSelectionExpression :: Text

The route selection expression for the API. For HTTP APIs, the routeSelectionExpression must be ${request.method} ${request.path}. If not provided, this will be the default for HTTP APIs. This property is required for WebSocket APIs.

name :: Text

The name of the API.

protocolType :: ProtocolType

The API protocol.

newApi

Arguments

:: Text
$sel:routeSelectionExpression:Api'

-> Text
$sel:name:Api'

-> ProtocolType
$sel:protocolType:Api'

-> Api

Create a value of Api with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:apiId:Api', api_apiId - The API ID.

$sel:disableExecuteApiEndpoint:Api', api_disableExecuteApiEndpoint - Specifies whether clients can invoke your API by using the default execute-api endpoint. By default, clients can invoke your API with the default https://{api_id}.execute-api.{region}.amazonaws.com endpoint. To require that clients use a custom domain name to invoke your API, disable the default endpoint.

$sel:apiEndpoint:Api', api_apiEndpoint - The URI of the API, of the form {api-id}.execute-api.{region}.amazonaws.com. The stage name is typically appended to this URI to form a complete path to a deployed API stage.

$sel:warnings:Api', api_warnings - The warning messages reported when failonwarnings is turned on during API import.

$sel:createdDate:Api', api_createdDate - The timestamp when the API was created.

$sel:version:Api', api_version - A version identifier for the API.

$sel:apiGatewayManaged:Api', api_apiGatewayManaged - Specifies whether an API is managed by API Gateway. You can't update or delete a managed API by using API Gateway. A managed API can be deleted only through the tooling or service that created it.

$sel:apiKeySelectionExpression:Api', api_apiKeySelectionExpression - An API key selection expression. Supported only for WebSocket APIs. See API Key Selection Expressions.

$sel:corsConfiguration:Api', api_corsConfiguration - A CORS configuration. Supported only for HTTP APIs.

$sel:importInfo:Api', api_importInfo - The validation information during API import. This may include particular properties of your OpenAPI definition which are ignored during import. Supported only for HTTP APIs.

$sel:disableSchemaValidation:Api', api_disableSchemaValidation - Avoid validating models when creating a deployment. Supported only for WebSocket APIs.

$sel:description:Api', api_description - The description of the API.

$sel:tags:Api', api_tags - A collection of tags associated with the API.

$sel:routeSelectionExpression:Api', api_routeSelectionExpression - The route selection expression for the API. For HTTP APIs, the routeSelectionExpression must be ${request.method} ${request.path}. If not provided, this will be the default for HTTP APIs. This property is required for WebSocket APIs.

$sel:name:Api', api_name - The name of the API.

$sel:protocolType:Api', api_protocolType - The API protocol.

api_apiId :: Lens' Api (Maybe Text)

The API ID.

api_disableExecuteApiEndpoint :: Lens' Api (Maybe Bool)

Specifies whether clients can invoke your API by using the default execute-api endpoint. By default, clients can invoke your API with the default https://{api_id}.execute-api.{region}.amazonaws.com endpoint. To require that clients use a custom domain name to invoke your API, disable the default endpoint.

api_apiEndpoint :: Lens' Api (Maybe Text)

The URI of the API, of the form {api-id}.execute-api.{region}.amazonaws.com. The stage name is typically appended to this URI to form a complete path to a deployed API stage.

api_warnings :: Lens' Api (Maybe [Text])

The warning messages reported when failonwarnings is turned on during API import.

api_createdDate :: Lens' Api (Maybe UTCTime)

The timestamp when the API was created.

api_version :: Lens' Api (Maybe Text)

A version identifier for the API.

api_apiGatewayManaged :: Lens' Api (Maybe Bool)

Specifies whether an API is managed by API Gateway. You can't update or delete a managed API by using API Gateway. A managed API can be deleted only through the tooling or service that created it.

api_apiKeySelectionExpression :: Lens' Api (Maybe Text)

An API key selection expression. Supported only for WebSocket APIs. See API Key Selection Expressions.

api_corsConfiguration :: Lens' Api (Maybe Cors)

A CORS configuration. Supported only for HTTP APIs.

api_importInfo :: Lens' Api (Maybe [Text])

The validation information during API import. This may include particular properties of your OpenAPI definition which are ignored during import. Supported only for HTTP APIs.

api_disableSchemaValidation :: Lens' Api (Maybe Bool)

Avoid validating models when creating a deployment. Supported only for WebSocket APIs.

api_description :: Lens' Api (Maybe Text)

The description of the API.

api_tags :: Lens' Api (Maybe (HashMap Text Text))

A collection of tags associated with the API.

api_routeSelectionExpression :: Lens' Api Text

The route selection expression for the API. For HTTP APIs, the routeSelectionExpression must be ${request.method} ${request.path}. If not provided, this will be the default for HTTP APIs. This property is required for WebSocket APIs.

api_name :: Lens' Api Text

The name of the API.

api_protocolType :: Lens' Api ProtocolType

The API protocol.