Amazonka.ApiGatewayV2.Types.DomainNameConfiguration

Description

Documentation

data DomainNameConfiguration

The domain name configuration.

See: newDomainNameConfiguration smart constructor.

Constructors

DomainNameConfiguration'

Fields

apiGatewayDomainName :: Maybe Text

A domain name for the API.

ownershipVerificationCertificateArn :: Maybe Text

The ARN of the public certificate issued by ACM to validate ownership of your custom domain. Only required when configuring mutual TLS and using an ACM imported or private CA certificate ARN as the regionalCertificateArn

certificateName :: Maybe Text

The user-friendly name of the certificate that will be used by the edge-optimized endpoint for this domain name.

hostedZoneId :: Maybe Text

The Amazon Route 53 Hosted Zone ID of the endpoint.

certificateArn :: Maybe Text

An AWS-managed certificate that will be used by the edge-optimized endpoint for this domain name. AWS Certificate Manager is the only supported source.

endpointType :: Maybe EndpointType

The endpoint type.

securityPolicy :: Maybe SecurityPolicy

The Transport Layer Security (TLS) version of the security policy for this domain name. The valid values are TLS_1_0 and TLS_1_2.

certificateUploadDate :: Maybe POSIX

The timestamp when the certificate that was used by edge-optimized endpoint for this domain name was uploaded.

domainNameStatusMessage :: Maybe Text

An optional text message containing detailed information about status of the domain name migration.

domainNameStatus :: Maybe DomainNameStatus

The status of the domain name migration. The valid values are AVAILABLE, UPDATING, PENDING_CERTIFICATE_REIMPORT, and PENDING_OWNERSHIP_VERIFICATION. If the status is UPDATING, the domain cannot be modified further until the existing operation is complete. If it is AVAILABLE, the domain can be updated.

newDomainNameConfiguration :: DomainNameConfiguration

Create a value of DomainNameConfiguration with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:apiGatewayDomainName:DomainNameConfiguration', domainNameConfiguration_apiGatewayDomainName - A domain name for the API.

$sel:ownershipVerificationCertificateArn:DomainNameConfiguration', domainNameConfiguration_ownershipVerificationCertificateArn - The ARN of the public certificate issued by ACM to validate ownership of your custom domain. Only required when configuring mutual TLS and using an ACM imported or private CA certificate ARN as the regionalCertificateArn

$sel:certificateName:DomainNameConfiguration', domainNameConfiguration_certificateName - The user-friendly name of the certificate that will be used by the edge-optimized endpoint for this domain name.

$sel:hostedZoneId:DomainNameConfiguration', domainNameConfiguration_hostedZoneId - The Amazon Route 53 Hosted Zone ID of the endpoint.

$sel:certificateArn:DomainNameConfiguration', domainNameConfiguration_certificateArn - An AWS-managed certificate that will be used by the edge-optimized endpoint for this domain name. AWS Certificate Manager is the only supported source.

$sel:endpointType:DomainNameConfiguration', domainNameConfiguration_endpointType - The endpoint type.

$sel:securityPolicy:DomainNameConfiguration', domainNameConfiguration_securityPolicy - The Transport Layer Security (TLS) version of the security policy for this domain name. The valid values are TLS_1_0 and TLS_1_2.

$sel:certificateUploadDate:DomainNameConfiguration', domainNameConfiguration_certificateUploadDate - The timestamp when the certificate that was used by edge-optimized endpoint for this domain name was uploaded.

$sel:domainNameStatusMessage:DomainNameConfiguration', domainNameConfiguration_domainNameStatusMessage - An optional text message containing detailed information about status of the domain name migration.

$sel:domainNameStatus:DomainNameConfiguration', domainNameConfiguration_domainNameStatus - The status of the domain name migration. The valid values are AVAILABLE, UPDATING, PENDING_CERTIFICATE_REIMPORT, and PENDING_OWNERSHIP_VERIFICATION. If the status is UPDATING, the domain cannot be modified further until the existing operation is complete. If it is AVAILABLE, the domain can be updated.

domainNameConfiguration_apiGatewayDomainName :: Lens' DomainNameConfiguration (Maybe Text)

A domain name for the API.

domainNameConfiguration_ownershipVerificationCertificateArn :: Lens' DomainNameConfiguration (Maybe Text)

The ARN of the public certificate issued by ACM to validate ownership of your custom domain. Only required when configuring mutual TLS and using an ACM imported or private CA certificate ARN as the regionalCertificateArn

domainNameConfiguration_certificateName :: Lens' DomainNameConfiguration (Maybe Text)

The user-friendly name of the certificate that will be used by the edge-optimized endpoint for this domain name.

domainNameConfiguration_hostedZoneId :: Lens' DomainNameConfiguration (Maybe Text)

The Amazon Route 53 Hosted Zone ID of the endpoint.

domainNameConfiguration_certificateArn :: Lens' DomainNameConfiguration (Maybe Text)

An AWS-managed certificate that will be used by the edge-optimized endpoint for this domain name. AWS Certificate Manager is the only supported source.

domainNameConfiguration_endpointType :: Lens' DomainNameConfiguration (Maybe EndpointType)

The endpoint type.

domainNameConfiguration_securityPolicy :: Lens' DomainNameConfiguration (Maybe SecurityPolicy)

The Transport Layer Security (TLS) version of the security policy for this domain name. The valid values are TLS_1_0 and TLS_1_2.

domainNameConfiguration_certificateUploadDate :: Lens' DomainNameConfiguration (Maybe UTCTime)

The timestamp when the certificate that was used by edge-optimized endpoint for this domain name was uploaded.

domainNameConfiguration_domainNameStatusMessage :: Lens' DomainNameConfiguration (Maybe Text)

An optional text message containing detailed information about status of the domain name migration.

domainNameConfiguration_domainNameStatus :: Lens' DomainNameConfiguration (Maybe DomainNameStatus)

The status of the domain name migration. The valid values are AVAILABLE, UPDATING, PENDING_CERTIFICATE_REIMPORT, and PENDING_OWNERSHIP_VERIFICATION. If the status is UPDATING, the domain cannot be modified further until the existing operation is complete. If it is AVAILABLE, the domain can be updated.