Amazonka.ApiGatewayV2.Types.IntegrationResponse

Description

Documentation

data IntegrationResponse

Represents an integration response.

See: newIntegrationResponse smart constructor.

Constructors

IntegrationResponse'

Fields

integrationResponseId :: Maybe Text

The integration response ID.

templateSelectionExpression :: Maybe Text

The template selection expressions for the integration response.

contentHandlingStrategy :: Maybe ContentHandlingStrategy

Supported only for WebSocket APIs. Specifies how to handle response payload content type conversions. Supported values are CONVERT_TO_BINARY and CONVERT_TO_TEXT, with the following behaviors:

CONVERT_TO_BINARY: Converts a response payload from a Base64-encoded string to the corresponding binary blob.

CONVERT_TO_TEXT: Converts a response payload from a binary blob to a Base64-encoded string.

If this property is not defined, the response payload will be passed through from the integration response to the route response or method response without modification.

responseTemplates :: Maybe (HashMap Text Text)

The collection of response templates for the integration response as a string-to-string map of key-value pairs. Response templates are represented as a key/value map, with a content-type as the key and a template as the value.

responseParameters :: Maybe (HashMap Text Text)

A key-value map specifying response parameters that are passed to the method response from the backend. The key is a method response header parameter name and the mapped value is an integration response header value, a static value enclosed within a pair of single quotes, or a JSON expression from the integration response body. The mapping key must match the pattern of method.response.header.{name}, where name is a valid and unique header name. The mapped non-static value must match the pattern of integration.response.header.{name} or integration.response.body.{JSON-expression}, where name is a valid and unique response header name and JSON-expression is a valid JSON expression without the $ prefix.

integrationResponseKey :: Text

The integration response key.

newIntegrationResponse

Arguments

:: Text
$sel:integrationResponseKey:IntegrationResponse'

-> IntegrationResponse

Create a value of IntegrationResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:integrationResponseId:IntegrationResponse', integrationResponse_integrationResponseId - The integration response ID.

$sel:templateSelectionExpression:IntegrationResponse', integrationResponse_templateSelectionExpression - The template selection expressions for the integration response.

$sel:contentHandlingStrategy:IntegrationResponse', integrationResponse_contentHandlingStrategy - Supported only for WebSocket APIs. Specifies how to handle response payload content type conversions. Supported values are CONVERT_TO_BINARY and CONVERT_TO_TEXT, with the following behaviors:

CONVERT_TO_BINARY: Converts a response payload from a Base64-encoded string to the corresponding binary blob.

CONVERT_TO_TEXT: Converts a response payload from a binary blob to a Base64-encoded string.

If this property is not defined, the response payload will be passed through from the integration response to the route response or method response without modification.

$sel:responseTemplates:IntegrationResponse', integrationResponse_responseTemplates - The collection of response templates for the integration response as a string-to-string map of key-value pairs. Response templates are represented as a key/value map, with a content-type as the key and a template as the value.

$sel:responseParameters:IntegrationResponse', integrationResponse_responseParameters - A key-value map specifying response parameters that are passed to the method response from the backend. The key is a method response header parameter name and the mapped value is an integration response header value, a static value enclosed within a pair of single quotes, or a JSON expression from the integration response body. The mapping key must match the pattern of method.response.header.{name}, where name is a valid and unique header name. The mapped non-static value must match the pattern of integration.response.header.{name} or integration.response.body.{JSON-expression}, where name is a valid and unique response header name and JSON-expression is a valid JSON expression without the $ prefix.

$sel:integrationResponseKey:IntegrationResponse', integrationResponse_integrationResponseKey - The integration response key.

integrationResponse_integrationResponseId :: Lens' IntegrationResponse (Maybe Text)

The integration response ID.

integrationResponse_templateSelectionExpression :: Lens' IntegrationResponse (Maybe Text)

The template selection expressions for the integration response.

integrationResponse_contentHandlingStrategy :: Lens' IntegrationResponse (Maybe ContentHandlingStrategy)

Supported only for WebSocket APIs. Specifies how to handle response payload content type conversions. Supported values are CONVERT_TO_BINARY and CONVERT_TO_TEXT, with the following behaviors:

CONVERT_TO_BINARY: Converts a response payload from a Base64-encoded string to the corresponding binary blob.

CONVERT_TO_TEXT: Converts a response payload from a binary blob to a Base64-encoded string.

If this property is not defined, the response payload will be passed through from the integration response to the route response or method response without modification.

integrationResponse_responseTemplates :: Lens' IntegrationResponse (Maybe (HashMap Text Text))

The collection of response templates for the integration response as a string-to-string map of key-value pairs. Response templates are represented as a key/value map, with a content-type as the key and a template as the value.

integrationResponse_responseParameters :: Lens' IntegrationResponse (Maybe (HashMap Text Text))

A key-value map specifying response parameters that are passed to the method response from the backend. The key is a method response header parameter name and the mapped value is an integration response header value, a static value enclosed within a pair of single quotes, or a JSON expression from the integration response body. The mapping key must match the pattern of method.response.header.{name}, where name is a valid and unique header name. The mapped non-static value must match the pattern of integration.response.header.{name} or integration.response.body.{JSON-expression}, where name is a valid and unique response header name and JSON-expression is a valid JSON expression without the $ prefix.

integrationResponse_integrationResponseKey :: Lens' IntegrationResponse Text

The integration response key.