Amazonka.ApiGatewayV2.Types.JWTConfiguration

Description

Documentation

data JWTConfiguration

Represents the configuration of a JWT authorizer. Required for the JWT authorizer type. Supported only for HTTP APIs.

See: newJWTConfiguration smart constructor.

Constructors

JWTConfiguration'

Fields

audience :: Maybe [Text]

A list of the intended recipients of the JWT. A valid JWT must provide an aud that matches at least one entry in this list. See RFC 7519. Supported only for HTTP APIs.

issuer :: Maybe Text

The base domain of the identity provider that issues JSON Web Tokens. For example, an Amazon Cognito user pool has the following format: https://cognito-idp.{region}.amazonaws.com/{userPoolId} . Required for the JWT authorizer type. Supported only for HTTP APIs.

newJWTConfiguration :: JWTConfiguration

Create a value of JWTConfiguration with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:audience:JWTConfiguration', jWTConfiguration_audience - A list of the intended recipients of the JWT. A valid JWT must provide an aud that matches at least one entry in this list. See RFC 7519. Supported only for HTTP APIs.

$sel:issuer:JWTConfiguration', jWTConfiguration_issuer - The base domain of the identity provider that issues JSON Web Tokens. For example, an Amazon Cognito user pool has the following format: https://cognito-idp.{region}.amazonaws.com/{userPoolId} . Required for the JWT authorizer type. Supported only for HTTP APIs.

jWTConfiguration_audience :: Lens' JWTConfiguration (Maybe [Text])

A list of the intended recipients of the JWT. A valid JWT must provide an aud that matches at least one entry in this list. See RFC 7519. Supported only for HTTP APIs.

jWTConfiguration_issuer :: Lens' JWTConfiguration (Maybe Text)

The base domain of the identity provider that issues JSON Web Tokens. For example, an Amazon Cognito user pool has the following format: https://cognito-idp.{region}.amazonaws.com/{userPoolId} . Required for the JWT authorizer type. Supported only for HTTP APIs.