Amazonka.ApiGatewayV2.Types.Route

Description

Documentation

data Route

Represents a route.

See: newRoute smart constructor.

Constructors

Route'

Fields

authorizationScopes :: Maybe [Text]

A list of authorization scopes configured on a route. The scopes are used with a JWT authorizer to authorize the method invocation. The authorization works by matching the route scopes against the scopes parsed from the access token in the incoming request. The method invocation is authorized if any route scope matches a claimed scope in the access token. Otherwise, the invocation is not authorized. When the route scope is configured, the client must provide an access token instead of an identity token for authorization purposes.

modelSelectionExpression :: Maybe Text

The model selection expression for the route. Supported only for WebSocket APIs.

requestModels :: Maybe (HashMap Text Text)

The request models for the route. Supported only for WebSocket APIs.

routeResponseSelectionExpression :: Maybe Text

The route response selection expression for the route. Supported only for WebSocket APIs.

requestParameters :: Maybe (HashMap Text ParameterConstraints)

The request parameters for the route. Supported only for WebSocket APIs.

routeId :: Maybe Text

The route ID.

authorizerId :: Maybe Text

The identifier of the Authorizer resource to be associated with this route. The authorizer identifier is generated by API Gateway when you created the authorizer.

operationName :: Maybe Text

The operation name for the route.

apiGatewayManaged :: Maybe Bool

Specifies whether a route is managed by API Gateway. If you created an API using quick create, the $default route is managed by API Gateway. You can't modify the $default route key.

authorizationType :: Maybe AuthorizationType

The authorization type for the route. For WebSocket APIs, valid values are NONE for open access, AWS_IAM for using AWS IAM permissions, and CUSTOM for using a Lambda authorizer For HTTP APIs, valid values are NONE for open access, JWT for using JSON Web Tokens, AWS_IAM for using AWS IAM permissions, and CUSTOM for using a Lambda authorizer.

apiKeyRequired :: Maybe Bool

Specifies whether an API key is required for this route. Supported only for WebSocket APIs.

target :: Maybe Text

The target for the route.

routeKey :: Text

The route key for the route.

newRoute

Arguments

:: Text
$sel:routeKey:Route'

-> Route

Create a value of Route with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:authorizationScopes:Route', route_authorizationScopes - A list of authorization scopes configured on a route. The scopes are used with a JWT authorizer to authorize the method invocation. The authorization works by matching the route scopes against the scopes parsed from the access token in the incoming request. The method invocation is authorized if any route scope matches a claimed scope in the access token. Otherwise, the invocation is not authorized. When the route scope is configured, the client must provide an access token instead of an identity token for authorization purposes.

$sel:modelSelectionExpression:Route', route_modelSelectionExpression - The model selection expression for the route. Supported only for WebSocket APIs.

$sel:requestModels:Route', route_requestModels - The request models for the route. Supported only for WebSocket APIs.

$sel:routeResponseSelectionExpression:Route', route_routeResponseSelectionExpression - The route response selection expression for the route. Supported only for WebSocket APIs.

$sel:requestParameters:Route', route_requestParameters - The request parameters for the route. Supported only for WebSocket APIs.

$sel:routeId:Route', route_routeId - The route ID.

$sel:authorizerId:Route', route_authorizerId - The identifier of the Authorizer resource to be associated with this route. The authorizer identifier is generated by API Gateway when you created the authorizer.

$sel:operationName:Route', route_operationName - The operation name for the route.

$sel:apiGatewayManaged:Route', route_apiGatewayManaged - Specifies whether a route is managed by API Gateway. If you created an API using quick create, the $default route is managed by API Gateway. You can't modify the $default route key.

$sel:authorizationType:Route', route_authorizationType - The authorization type for the route. For WebSocket APIs, valid values are NONE for open access, AWS_IAM for using AWS IAM permissions, and CUSTOM for using a Lambda authorizer For HTTP APIs, valid values are NONE for open access, JWT for using JSON Web Tokens, AWS_IAM for using AWS IAM permissions, and CUSTOM for using a Lambda authorizer.

$sel:apiKeyRequired:Route', route_apiKeyRequired - Specifies whether an API key is required for this route. Supported only for WebSocket APIs.

$sel:target:Route', route_target - The target for the route.

$sel:routeKey:Route', route_routeKey - The route key for the route.

route_authorizationScopes :: Lens' Route (Maybe [Text])

A list of authorization scopes configured on a route. The scopes are used with a JWT authorizer to authorize the method invocation. The authorization works by matching the route scopes against the scopes parsed from the access token in the incoming request. The method invocation is authorized if any route scope matches a claimed scope in the access token. Otherwise, the invocation is not authorized. When the route scope is configured, the client must provide an access token instead of an identity token for authorization purposes.

route_modelSelectionExpression :: Lens' Route (Maybe Text)

The model selection expression for the route. Supported only for WebSocket APIs.

route_requestModels :: Lens' Route (Maybe (HashMap Text Text))

The request models for the route. Supported only for WebSocket APIs.

route_routeResponseSelectionExpression :: Lens' Route (Maybe Text)

The route response selection expression for the route. Supported only for WebSocket APIs.

route_requestParameters :: Lens' Route (Maybe (HashMap Text ParameterConstraints))

The request parameters for the route. Supported only for WebSocket APIs.

route_routeId :: Lens' Route (Maybe Text)

The route ID.

route_authorizerId :: Lens' Route (Maybe Text)

The identifier of the Authorizer resource to be associated with this route. The authorizer identifier is generated by API Gateway when you created the authorizer.

route_operationName :: Lens' Route (Maybe Text)

The operation name for the route.

route_apiGatewayManaged :: Lens' Route (Maybe Bool)

Specifies whether a route is managed by API Gateway. If you created an API using quick create, the $default route is managed by API Gateway. You can't modify the $default route key.

route_authorizationType :: Lens' Route (Maybe AuthorizationType)

The authorization type for the route. For WebSocket APIs, valid values are NONE for open access, AWS_IAM for using AWS IAM permissions, and CUSTOM for using a Lambda authorizer For HTTP APIs, valid values are NONE for open access, JWT for using JSON Web Tokens, AWS_IAM for using AWS IAM permissions, and CUSTOM for using a Lambda authorizer.

route_apiKeyRequired :: Lens' Route (Maybe Bool)

Specifies whether an API key is required for this route. Supported only for WebSocket APIs.

route_target :: Lens' Route (Maybe Text)

The target for the route.

route_routeKey :: Lens' Route Text

The route key for the route.