Amazonka.ApiGatewayV2.UpdateApi

Description

Updates an Api resource.

Creating a Request

data UpdateApi

Updates an Api.

See: newUpdateApi smart constructor.

Constructors

UpdateApi'

Fields

credentialsArn :: Maybe Text

This property is part of quick create. It specifies the credentials required for the integration, if any. For a Lambda integration, three options are available. To specify an IAM Role for API Gateway to assume, use the role's Amazon Resource Name (ARN). To require that the caller's identity be passed through from the request, specify arn:aws:iam::*:user/*. To use resource-based permissions on supported AWS services, don't specify this parameter. Currently, this property is not used for HTTP integrations. If provided, this value replaces the credentials associated with the quick create integration. Supported only for HTTP APIs.

disableExecuteApiEndpoint :: Maybe Bool

Specifies whether clients can invoke your API by using the default execute-api endpoint. By default, clients can invoke your API with the default https://{api_id}.execute-api.{region}.amazonaws.com endpoint. To require that clients use a custom domain name to invoke your API, disable the default endpoint.

name :: Maybe Text

The name of the API.

version :: Maybe Text

A version identifier for the API.

apiKeySelectionExpression :: Maybe Text

An API key selection expression. Supported only for WebSocket APIs. See API Key Selection Expressions.

corsConfiguration :: Maybe Cors

A CORS configuration. Supported only for HTTP APIs.

routeSelectionExpression :: Maybe Text

The route selection expression for the API. For HTTP APIs, the routeSelectionExpression must be ${request.method} ${request.path}. If not provided, this will be the default for HTTP APIs. This property is required for WebSocket APIs.

disableSchemaValidation :: Maybe Bool

Avoid validating models when creating a deployment. Supported only for WebSocket APIs.

description :: Maybe Text

The description of the API.

routeKey :: Maybe Text

This property is part of quick create. If not specified, the route created using quick create is kept. Otherwise, this value replaces the route key of the quick create route. Additional routes may still be added after the API is updated. Supported only for HTTP APIs.

target :: Maybe Text

This property is part of quick create. For HTTP integrations, specify a fully qualified URL. For Lambda integrations, specify a function ARN. The type of the integration will be HTTP_PROXY or AWS_PROXY, respectively. The value provided updates the integration URI and integration type. You can update a quick-created target, but you can't remove it from an API. Supported only for HTTP APIs.

apiId :: Text

The API identifier.

newUpdateApi

Arguments

:: Text
$sel:apiId:UpdateApi'

-> UpdateApi

Create a value of UpdateApi with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:credentialsArn:UpdateApi', updateApi_credentialsArn - This property is part of quick create. It specifies the credentials required for the integration, if any. For a Lambda integration, three options are available. To specify an IAM Role for API Gateway to assume, use the role's Amazon Resource Name (ARN). To require that the caller's identity be passed through from the request, specify arn:aws:iam::*:user/*. To use resource-based permissions on supported AWS services, don't specify this parameter. Currently, this property is not used for HTTP integrations. If provided, this value replaces the credentials associated with the quick create integration. Supported only for HTTP APIs.

$sel:disableExecuteApiEndpoint:UpdateApi', updateApi_disableExecuteApiEndpoint - Specifies whether clients can invoke your API by using the default execute-api endpoint. By default, clients can invoke your API with the default https://{api_id}.execute-api.{region}.amazonaws.com endpoint. To require that clients use a custom domain name to invoke your API, disable the default endpoint.

$sel:name:UpdateApi', updateApi_name - The name of the API.

$sel:version:UpdateApi', updateApi_version - A version identifier for the API.

$sel:apiKeySelectionExpression:UpdateApi', updateApi_apiKeySelectionExpression - An API key selection expression. Supported only for WebSocket APIs. See API Key Selection Expressions.

$sel:corsConfiguration:UpdateApi', updateApi_corsConfiguration - A CORS configuration. Supported only for HTTP APIs.

$sel:routeSelectionExpression:UpdateApi', updateApi_routeSelectionExpression - The route selection expression for the API. For HTTP APIs, the routeSelectionExpression must be ${request.method} ${request.path}. If not provided, this will be the default for HTTP APIs. This property is required for WebSocket APIs.

$sel:disableSchemaValidation:UpdateApi', updateApi_disableSchemaValidation - Avoid validating models when creating a deployment. Supported only for WebSocket APIs.

$sel:description:UpdateApi', updateApi_description - The description of the API.

$sel:routeKey:UpdateApi', updateApi_routeKey - This property is part of quick create. If not specified, the route created using quick create is kept. Otherwise, this value replaces the route key of the quick create route. Additional routes may still be added after the API is updated. Supported only for HTTP APIs.

$sel:target:UpdateApi', updateApi_target - This property is part of quick create. For HTTP integrations, specify a fully qualified URL. For Lambda integrations, specify a function ARN. The type of the integration will be HTTP_PROXY or AWS_PROXY, respectively. The value provided updates the integration URI and integration type. You can update a quick-created target, but you can't remove it from an API. Supported only for HTTP APIs.

$sel:apiId:UpdateApi', updateApi_apiId - The API identifier.

Request Lenses

updateApi_credentialsArn :: Lens' UpdateApi (Maybe Text)

This property is part of quick create. It specifies the credentials required for the integration, if any. For a Lambda integration, three options are available. To specify an IAM Role for API Gateway to assume, use the role's Amazon Resource Name (ARN). To require that the caller's identity be passed through from the request, specify arn:aws:iam::*:user/*. To use resource-based permissions on supported AWS services, don't specify this parameter. Currently, this property is not used for HTTP integrations. If provided, this value replaces the credentials associated with the quick create integration. Supported only for HTTP APIs.

updateApi_disableExecuteApiEndpoint :: Lens' UpdateApi (Maybe Bool)

Specifies whether clients can invoke your API by using the default execute-api endpoint. By default, clients can invoke your API with the default https://{api_id}.execute-api.{region}.amazonaws.com endpoint. To require that clients use a custom domain name to invoke your API, disable the default endpoint.

updateApi_name :: Lens' UpdateApi (Maybe Text)

The name of the API.

updateApi_version :: Lens' UpdateApi (Maybe Text)

A version identifier for the API.

updateApi_apiKeySelectionExpression :: Lens' UpdateApi (Maybe Text)

An API key selection expression. Supported only for WebSocket APIs. See API Key Selection Expressions.

updateApi_corsConfiguration :: Lens' UpdateApi (Maybe Cors)

A CORS configuration. Supported only for HTTP APIs.

updateApi_routeSelectionExpression :: Lens' UpdateApi (Maybe Text)

The route selection expression for the API. For HTTP APIs, the routeSelectionExpression must be ${request.method} ${request.path}. If not provided, this will be the default for HTTP APIs. This property is required for WebSocket APIs.

updateApi_disableSchemaValidation :: Lens' UpdateApi (Maybe Bool)

Avoid validating models when creating a deployment. Supported only for WebSocket APIs.

updateApi_description :: Lens' UpdateApi (Maybe Text)

The description of the API.

updateApi_routeKey :: Lens' UpdateApi (Maybe Text)

This property is part of quick create. If not specified, the route created using quick create is kept. Otherwise, this value replaces the route key of the quick create route. Additional routes may still be added after the API is updated. Supported only for HTTP APIs.

updateApi_target :: Lens' UpdateApi (Maybe Text)

This property is part of quick create. For HTTP integrations, specify a fully qualified URL. For Lambda integrations, specify a function ARN. The type of the integration will be HTTP_PROXY or AWS_PROXY, respectively. The value provided updates the integration URI and integration type. You can update a quick-created target, but you can't remove it from an API. Supported only for HTTP APIs.

updateApi_apiId :: Lens' UpdateApi Text

The API identifier.

Destructuring the Response

data UpdateApiResponse

See: newUpdateApiResponse smart constructor.

Constructors

UpdateApiResponse'

Fields

apiId :: Maybe Text

The API ID.

disableExecuteApiEndpoint :: Maybe Bool

Specifies whether clients can invoke your API by using the default execute-api endpoint. By default, clients can invoke your API with the default https://{api_id}.execute-api.{region}.amazonaws.com endpoint. To require that clients use a custom domain name to invoke your API, disable the default endpoint.

apiEndpoint :: Maybe Text

The URI of the API, of the form {api-id}.execute-api.{region}.amazonaws.com. The stage name is typically appended to this URI to form a complete path to a deployed API stage.

warnings :: Maybe [Text]

The warning messages reported when failonwarnings is turned on during API import.

createdDate :: Maybe POSIX

The timestamp when the API was created.

name :: Maybe Text

The name of the API.

version :: Maybe Text

A version identifier for the API.

apiGatewayManaged :: Maybe Bool

Specifies whether an API is managed by API Gateway. You can't update or delete a managed API by using API Gateway. A managed API can be deleted only through the tooling or service that created it.

apiKeySelectionExpression :: Maybe Text

An API key selection expression. Supported only for WebSocket APIs. See API Key Selection Expressions.

corsConfiguration :: Maybe Cors

A CORS configuration. Supported only for HTTP APIs.

routeSelectionExpression :: Maybe Text

The route selection expression for the API. For HTTP APIs, the routeSelectionExpression must be ${request.method} ${request.path}. If not provided, this will be the default for HTTP APIs. This property is required for WebSocket APIs.

importInfo :: Maybe [Text]

The validation information during API import. This may include particular properties of your OpenAPI definition which are ignored during import. Supported only for HTTP APIs.

disableSchemaValidation :: Maybe Bool

Avoid validating models when creating a deployment. Supported only for WebSocket APIs.

description :: Maybe Text

The description of the API.

protocolType :: Maybe ProtocolType

The API protocol.

tags :: Maybe (HashMap Text Text)

A collection of tags associated with the API.

httpStatus :: Int

The response's http status code.

newUpdateApiResponse

Arguments

:: Int
$sel:httpStatus:UpdateApiResponse'

-> UpdateApiResponse

Create a value of UpdateApiResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:apiId:UpdateApi', updateApiResponse_apiId - The API ID.

$sel:disableExecuteApiEndpoint:UpdateApi', updateApiResponse_disableExecuteApiEndpoint - Specifies whether clients can invoke your API by using the default execute-api endpoint. By default, clients can invoke your API with the default https://{api_id}.execute-api.{region}.amazonaws.com endpoint. To require that clients use a custom domain name to invoke your API, disable the default endpoint.

$sel:apiEndpoint:UpdateApiResponse', updateApiResponse_apiEndpoint - The URI of the API, of the form {api-id}.execute-api.{region}.amazonaws.com. The stage name is typically appended to this URI to form a complete path to a deployed API stage.

$sel:warnings:UpdateApiResponse', updateApiResponse_warnings - The warning messages reported when failonwarnings is turned on during API import.

$sel:createdDate:UpdateApiResponse', updateApiResponse_createdDate - The timestamp when the API was created.

$sel:name:UpdateApi', updateApiResponse_name - The name of the API.

$sel:version:UpdateApi', updateApiResponse_version - A version identifier for the API.

$sel:apiGatewayManaged:UpdateApiResponse', updateApiResponse_apiGatewayManaged - Specifies whether an API is managed by API Gateway. You can't update or delete a managed API by using API Gateway. A managed API can be deleted only through the tooling or service that created it.

$sel:apiKeySelectionExpression:UpdateApi', updateApiResponse_apiKeySelectionExpression - An API key selection expression. Supported only for WebSocket APIs. See API Key Selection Expressions.

$sel:corsConfiguration:UpdateApi', updateApiResponse_corsConfiguration - A CORS configuration. Supported only for HTTP APIs.

$sel:routeSelectionExpression:UpdateApi', updateApiResponse_routeSelectionExpression - The route selection expression for the API. For HTTP APIs, the routeSelectionExpression must be ${request.method} ${request.path}. If not provided, this will be the default for HTTP APIs. This property is required for WebSocket APIs.

$sel:importInfo:UpdateApiResponse', updateApiResponse_importInfo - The validation information during API import. This may include particular properties of your OpenAPI definition which are ignored during import. Supported only for HTTP APIs.

$sel:disableSchemaValidation:UpdateApi', updateApiResponse_disableSchemaValidation - Avoid validating models when creating a deployment. Supported only for WebSocket APIs.

$sel:description:UpdateApi', updateApiResponse_description - The description of the API.

$sel:protocolType:UpdateApiResponse', updateApiResponse_protocolType - The API protocol.

$sel:tags:UpdateApiResponse', updateApiResponse_tags - A collection of tags associated with the API.

$sel:httpStatus:UpdateApiResponse', updateApiResponse_httpStatus - The response's http status code.

Response Lenses

updateApiResponse_apiId :: Lens' UpdateApiResponse (Maybe Text)

The API ID.

updateApiResponse_disableExecuteApiEndpoint :: Lens' UpdateApiResponse (Maybe Bool)

Specifies whether clients can invoke your API by using the default execute-api endpoint. By default, clients can invoke your API with the default https://{api_id}.execute-api.{region}.amazonaws.com endpoint. To require that clients use a custom domain name to invoke your API, disable the default endpoint.

updateApiResponse_apiEndpoint :: Lens' UpdateApiResponse (Maybe Text)

The URI of the API, of the form {api-id}.execute-api.{region}.amazonaws.com. The stage name is typically appended to this URI to form a complete path to a deployed API stage.

updateApiResponse_warnings :: Lens' UpdateApiResponse (Maybe [Text])

The warning messages reported when failonwarnings is turned on during API import.

updateApiResponse_createdDate :: Lens' UpdateApiResponse (Maybe UTCTime)

The timestamp when the API was created.

updateApiResponse_name :: Lens' UpdateApiResponse (Maybe Text)

The name of the API.

updateApiResponse_version :: Lens' UpdateApiResponse (Maybe Text)

A version identifier for the API.

updateApiResponse_apiGatewayManaged :: Lens' UpdateApiResponse (Maybe Bool)

Specifies whether an API is managed by API Gateway. You can't update or delete a managed API by using API Gateway. A managed API can be deleted only through the tooling or service that created it.

updateApiResponse_apiKeySelectionExpression :: Lens' UpdateApiResponse (Maybe Text)

An API key selection expression. Supported only for WebSocket APIs. See API Key Selection Expressions.

updateApiResponse_corsConfiguration :: Lens' UpdateApiResponse (Maybe Cors)

A CORS configuration. Supported only for HTTP APIs.

updateApiResponse_routeSelectionExpression :: Lens' UpdateApiResponse (Maybe Text)

The route selection expression for the API. For HTTP APIs, the routeSelectionExpression must be ${request.method} ${request.path}. If not provided, this will be the default for HTTP APIs. This property is required for WebSocket APIs.

updateApiResponse_importInfo :: Lens' UpdateApiResponse (Maybe [Text])

The validation information during API import. This may include particular properties of your OpenAPI definition which are ignored during import. Supported only for HTTP APIs.

updateApiResponse_disableSchemaValidation :: Lens' UpdateApiResponse (Maybe Bool)

Avoid validating models when creating a deployment. Supported only for WebSocket APIs.

updateApiResponse_description :: Lens' UpdateApiResponse (Maybe Text)

The description of the API.

updateApiResponse_protocolType :: Lens' UpdateApiResponse (Maybe ProtocolType)

The API protocol.

updateApiResponse_tags :: Lens' UpdateApiResponse (Maybe (HashMap Text Text))

A collection of tags associated with the API.

updateApiResponse_httpStatus :: Lens' UpdateApiResponse Int

The response's http status code.