Amazonka.ApiGatewayV2.UpdateStage

Description

Updates a Stage.

Creating a Request

data UpdateStage

Updates a Stage.

See: newUpdateStage smart constructor.

Constructors

UpdateStage'

Fields

deploymentId :: Maybe Text

The deployment identifier for the API stage. Can't be updated if autoDeploy is enabled.

routeSettings :: Maybe (HashMap Text RouteSettings)

Route settings for the stage.

accessLogSettings :: Maybe AccessLogSettings

Settings for logging access in this stage.

clientCertificateId :: Maybe Text

The identifier of a client certificate for a Stage.

stageVariables :: Maybe (HashMap Text Text)

A map that defines the stage variables for a Stage. Variable names can have alphanumeric and underscore characters, and the values must match [A-Za-z0-9-._~:/?#&=,]+.

autoDeploy :: Maybe Bool

Specifies whether updates to an API automatically trigger a new deployment. The default value is false.

defaultRouteSettings :: Maybe RouteSettings

The default route settings for the stage.

description :: Maybe Text

The description for the API stage.

stageName :: Text

The stage name. Stage names can contain only alphanumeric characters, hyphens, and underscores, or be $default. Maximum length is 128 characters.

apiId :: Text

The API identifier.

newUpdateStage

Arguments

:: Text
$sel:stageName:UpdateStage'

-> Text
$sel:apiId:UpdateStage'

-> UpdateStage

Create a value of UpdateStage with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:deploymentId:UpdateStage', updateStage_deploymentId - The deployment identifier for the API stage. Can't be updated if autoDeploy is enabled.

$sel:routeSettings:UpdateStage', updateStage_routeSettings - Route settings for the stage.

$sel:accessLogSettings:UpdateStage', updateStage_accessLogSettings - Settings for logging access in this stage.

$sel:clientCertificateId:UpdateStage', updateStage_clientCertificateId - The identifier of a client certificate for a Stage.

$sel:stageVariables:UpdateStage', updateStage_stageVariables - A map that defines the stage variables for a Stage. Variable names can have alphanumeric and underscore characters, and the values must match [A-Za-z0-9-._~:/?#&=,]+.

$sel:autoDeploy:UpdateStage', updateStage_autoDeploy - Specifies whether updates to an API automatically trigger a new deployment. The default value is false.

$sel:defaultRouteSettings:UpdateStage', updateStage_defaultRouteSettings - The default route settings for the stage.

$sel:description:UpdateStage', updateStage_description - The description for the API stage.

$sel:stageName:UpdateStage', updateStage_stageName - The stage name. Stage names can contain only alphanumeric characters, hyphens, and underscores, or be $default. Maximum length is 128 characters.

$sel:apiId:UpdateStage', updateStage_apiId - The API identifier.

Request Lenses

updateStage_deploymentId :: Lens' UpdateStage (Maybe Text)

The deployment identifier for the API stage. Can't be updated if autoDeploy is enabled.

updateStage_routeSettings :: Lens' UpdateStage (Maybe (HashMap Text RouteSettings))

Route settings for the stage.

updateStage_accessLogSettings :: Lens' UpdateStage (Maybe AccessLogSettings)

Settings for logging access in this stage.

updateStage_clientCertificateId :: Lens' UpdateStage (Maybe Text)

The identifier of a client certificate for a Stage.

updateStage_stageVariables :: Lens' UpdateStage (Maybe (HashMap Text Text))

A map that defines the stage variables for a Stage. Variable names can have alphanumeric and underscore characters, and the values must match [A-Za-z0-9-._~:/?#&=,]+.

updateStage_autoDeploy :: Lens' UpdateStage (Maybe Bool)

Specifies whether updates to an API automatically trigger a new deployment. The default value is false.

updateStage_defaultRouteSettings :: Lens' UpdateStage (Maybe RouteSettings)

The default route settings for the stage.

updateStage_description :: Lens' UpdateStage (Maybe Text)

The description for the API stage.

updateStage_stageName :: Lens' UpdateStage Text

The stage name. Stage names can contain only alphanumeric characters, hyphens, and underscores, or be $default. Maximum length is 128 characters.

updateStage_apiId :: Lens' UpdateStage Text

The API identifier.

Destructuring the Response

data UpdateStageResponse

See: newUpdateStageResponse smart constructor.

Constructors

UpdateStageResponse'

Fields

lastDeploymentStatusMessage :: Maybe Text

Describes the status of the last deployment of a stage. Supported only for stages with autoDeploy enabled.

deploymentId :: Maybe Text

The identifier of the Deployment that the Stage is associated with. Can't be updated if autoDeploy is enabled.

routeSettings :: Maybe (HashMap Text RouteSettings)

Route settings for the stage, by routeKey.

accessLogSettings :: Maybe AccessLogSettings

Settings for logging access in this stage.

clientCertificateId :: Maybe Text

The identifier of a client certificate for a Stage. Supported only for WebSocket APIs.

stageVariables :: Maybe (HashMap Text Text)

A map that defines the stage variables for a stage resource. Variable names can have alphanumeric and underscore characters, and the values must match [A-Za-z0-9-._~:/?#&=,]+.

autoDeploy :: Maybe Bool

Specifies whether updates to an API automatically trigger a new deployment. The default value is false.

createdDate :: Maybe POSIX

The timestamp when the stage was created.

defaultRouteSettings :: Maybe RouteSettings

Default route settings for the stage.

apiGatewayManaged :: Maybe Bool

Specifies whether a stage is managed by API Gateway. If you created an API using quick create, the $default stage is managed by API Gateway. You can't modify the $default stage.

stageName :: Maybe Text

The name of the stage.

lastUpdatedDate :: Maybe POSIX

The timestamp when the stage was last updated.

description :: Maybe Text

The description of the stage.

tags :: Maybe (HashMap Text Text)

The collection of tags. Each tag element is associated with a given resource.

httpStatus :: Int

The response's http status code.

newUpdateStageResponse

Arguments

:: Int
$sel:httpStatus:UpdateStageResponse'

-> UpdateStageResponse

Create a value of UpdateStageResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:lastDeploymentStatusMessage:UpdateStageResponse', updateStageResponse_lastDeploymentStatusMessage - Describes the status of the last deployment of a stage. Supported only for stages with autoDeploy enabled.

$sel:deploymentId:UpdateStage', updateStageResponse_deploymentId - The identifier of the Deployment that the Stage is associated with. Can't be updated if autoDeploy is enabled.

$sel:routeSettings:UpdateStage', updateStageResponse_routeSettings - Route settings for the stage, by routeKey.

$sel:accessLogSettings:UpdateStage', updateStageResponse_accessLogSettings - Settings for logging access in this stage.

$sel:clientCertificateId:UpdateStage', updateStageResponse_clientCertificateId - The identifier of a client certificate for a Stage. Supported only for WebSocket APIs.

$sel:stageVariables:UpdateStage', updateStageResponse_stageVariables - A map that defines the stage variables for a stage resource. Variable names can have alphanumeric and underscore characters, and the values must match [A-Za-z0-9-._~:/?#&=,]+.

$sel:autoDeploy:UpdateStage', updateStageResponse_autoDeploy - Specifies whether updates to an API automatically trigger a new deployment. The default value is false.

$sel:createdDate:UpdateStageResponse', updateStageResponse_createdDate - The timestamp when the stage was created.

$sel:defaultRouteSettings:UpdateStage', updateStageResponse_defaultRouteSettings - Default route settings for the stage.

$sel:apiGatewayManaged:UpdateStageResponse', updateStageResponse_apiGatewayManaged - Specifies whether a stage is managed by API Gateway. If you created an API using quick create, the $default stage is managed by API Gateway. You can't modify the $default stage.

$sel:stageName:UpdateStage', updateStageResponse_stageName - The name of the stage.

$sel:lastUpdatedDate:UpdateStageResponse', updateStageResponse_lastUpdatedDate - The timestamp when the stage was last updated.

$sel:description:UpdateStage', updateStageResponse_description - The description of the stage.

$sel:tags:UpdateStageResponse', updateStageResponse_tags - The collection of tags. Each tag element is associated with a given resource.

$sel:httpStatus:UpdateStageResponse', updateStageResponse_httpStatus - The response's http status code.

Response Lenses

updateStageResponse_lastDeploymentStatusMessage :: Lens' UpdateStageResponse (Maybe Text)

Describes the status of the last deployment of a stage. Supported only for stages with autoDeploy enabled.

updateStageResponse_deploymentId :: Lens' UpdateStageResponse (Maybe Text)

The identifier of the Deployment that the Stage is associated with. Can't be updated if autoDeploy is enabled.

updateStageResponse_routeSettings :: Lens' UpdateStageResponse (Maybe (HashMap Text RouteSettings))

Route settings for the stage, by routeKey.

updateStageResponse_accessLogSettings :: Lens' UpdateStageResponse (Maybe AccessLogSettings)

Settings for logging access in this stage.

updateStageResponse_clientCertificateId :: Lens' UpdateStageResponse (Maybe Text)

The identifier of a client certificate for a Stage. Supported only for WebSocket APIs.

updateStageResponse_stageVariables :: Lens' UpdateStageResponse (Maybe (HashMap Text Text))

A map that defines the stage variables for a stage resource. Variable names can have alphanumeric and underscore characters, and the values must match [A-Za-z0-9-._~:/?#&=,]+.

updateStageResponse_autoDeploy :: Lens' UpdateStageResponse (Maybe Bool)

Specifies whether updates to an API automatically trigger a new deployment. The default value is false.

updateStageResponse_createdDate :: Lens' UpdateStageResponse (Maybe UTCTime)

The timestamp when the stage was created.

updateStageResponse_defaultRouteSettings :: Lens' UpdateStageResponse (Maybe RouteSettings)

Default route settings for the stage.

updateStageResponse_apiGatewayManaged :: Lens' UpdateStageResponse (Maybe Bool)

Specifies whether a stage is managed by API Gateway. If you created an API using quick create, the $default stage is managed by API Gateway. You can't modify the $default stage.

updateStageResponse_stageName :: Lens' UpdateStageResponse (Maybe Text)

The name of the stage.

updateStageResponse_lastUpdatedDate :: Lens' UpdateStageResponse (Maybe UTCTime)

The timestamp when the stage was last updated.

updateStageResponse_description :: Lens' UpdateStageResponse (Maybe Text)

The description of the stage.

updateStageResponse_tags :: Lens' UpdateStageResponse (Maybe (HashMap Text Text))

The collection of tags. Each tag element is associated with a given resource.

updateStageResponse_httpStatus :: Lens' UpdateStageResponse Int

The response's http status code.