Amazonka.ApiGatewayV2.UpdateVpcLink

Description

Updates a VPC link.

Creating a Request

data UpdateVpcLink

Updates a VPC link.

See: newUpdateVpcLink smart constructor.

Constructors

UpdateVpcLink'

Fields

name :: Maybe Text

The name of the VPC link.

vpcLinkId :: Text

The ID of the VPC link.

newUpdateVpcLink

Arguments

:: Text
$sel:vpcLinkId:UpdateVpcLink'

-> UpdateVpcLink

Create a value of UpdateVpcLink with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:name:UpdateVpcLink', updateVpcLink_name - The name of the VPC link.

$sel:vpcLinkId:UpdateVpcLink', updateVpcLink_vpcLinkId - The ID of the VPC link.

Request Lenses

updateVpcLink_name :: Lens' UpdateVpcLink (Maybe Text)

The name of the VPC link.

updateVpcLink_vpcLinkId :: Lens' UpdateVpcLink Text

The ID of the VPC link.

Destructuring the Response

data UpdateVpcLinkResponse

See: newUpdateVpcLinkResponse smart constructor.

Constructors

UpdateVpcLinkResponse'

Fields

securityGroupIds :: Maybe [Text]

A list of security group IDs for the VPC link.

subnetIds :: Maybe [Text]

A list of subnet IDs to include in the VPC link.

vpcLinkId :: Maybe Text

The ID of the VPC link.

createdDate :: Maybe POSIX

The timestamp when the VPC link was created.

vpcLinkVersion :: Maybe VpcLinkVersion

The version of the VPC link.

name :: Maybe Text

The name of the VPC link.

vpcLinkStatusMessage :: Maybe Text

A message summarizing the cause of the status of the VPC link.

tags :: Maybe (HashMap Text Text)

Tags for the VPC link.

vpcLinkStatus :: Maybe VpcLinkStatus

The status of the VPC link.

httpStatus :: Int

The response's http status code.

newUpdateVpcLinkResponse

Arguments

:: Int
$sel:httpStatus:UpdateVpcLinkResponse'

-> UpdateVpcLinkResponse

Create a value of UpdateVpcLinkResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:securityGroupIds:UpdateVpcLinkResponse', updateVpcLinkResponse_securityGroupIds - A list of security group IDs for the VPC link.

$sel:subnetIds:UpdateVpcLinkResponse', updateVpcLinkResponse_subnetIds - A list of subnet IDs to include in the VPC link.

$sel:vpcLinkId:UpdateVpcLink', updateVpcLinkResponse_vpcLinkId - The ID of the VPC link.

$sel:createdDate:UpdateVpcLinkResponse', updateVpcLinkResponse_createdDate - The timestamp when the VPC link was created.

$sel:vpcLinkVersion:UpdateVpcLinkResponse', updateVpcLinkResponse_vpcLinkVersion - The version of the VPC link.

$sel:name:UpdateVpcLink', updateVpcLinkResponse_name - The name of the VPC link.

$sel:vpcLinkStatusMessage:UpdateVpcLinkResponse', updateVpcLinkResponse_vpcLinkStatusMessage - A message summarizing the cause of the status of the VPC link.

$sel:tags:UpdateVpcLinkResponse', updateVpcLinkResponse_tags - Tags for the VPC link.

$sel:vpcLinkStatus:UpdateVpcLinkResponse', updateVpcLinkResponse_vpcLinkStatus - The status of the VPC link.

$sel:httpStatus:UpdateVpcLinkResponse', updateVpcLinkResponse_httpStatus - The response's http status code.

Response Lenses

updateVpcLinkResponse_securityGroupIds :: Lens' UpdateVpcLinkResponse (Maybe [Text])

A list of security group IDs for the VPC link.

updateVpcLinkResponse_subnetIds :: Lens' UpdateVpcLinkResponse (Maybe [Text])

A list of subnet IDs to include in the VPC link.

updateVpcLinkResponse_vpcLinkId :: Lens' UpdateVpcLinkResponse (Maybe Text)

The ID of the VPC link.

updateVpcLinkResponse_createdDate :: Lens' UpdateVpcLinkResponse (Maybe UTCTime)

The timestamp when the VPC link was created.

updateVpcLinkResponse_vpcLinkVersion :: Lens' UpdateVpcLinkResponse (Maybe VpcLinkVersion)

The version of the VPC link.

updateVpcLinkResponse_name :: Lens' UpdateVpcLinkResponse (Maybe Text)

The name of the VPC link.

updateVpcLinkResponse_vpcLinkStatusMessage :: Lens' UpdateVpcLinkResponse (Maybe Text)

A message summarizing the cause of the status of the VPC link.

updateVpcLinkResponse_tags :: Lens' UpdateVpcLinkResponse (Maybe (HashMap Text Text))

Tags for the VPC link.

updateVpcLinkResponse_vpcLinkStatus :: Lens' UpdateVpcLinkResponse (Maybe VpcLinkStatus)

The status of the VPC link.

updateVpcLinkResponse_httpStatus :: Lens' UpdateVpcLinkResponse Int

The response's http status code.