Amazonka.AppConfig.CreateConfigurationProfile

Description

Information that enables AppConfig to access the configuration source. Valid configuration sources include Systems Manager (SSM) documents, SSM Parameter Store parameters, and Amazon S3 objects. A configuration profile includes the following information.

The Uri location of the configuration data.

The AWS Identity and Access Management (IAM) role that provides access to the configuration data.

A validator for the configuration data. Available validators include either a JSON Schema or an AWS Lambda function.

For more information, see Create a Configuration and a Configuration Profile in the AWS AppConfig User Guide.

Creating a Request

data CreateConfigurationProfile

See: newCreateConfigurationProfile smart constructor.

Constructors

CreateConfigurationProfile'

Fields

retrievalRoleArn :: Maybe Text

The ARN of an IAM role with permission to access the configuration at the specified LocationUri.

validators :: Maybe [Validator]

A list of methods for validating the configuration.

description :: Maybe Text

A description of the configuration profile.

tags :: Maybe (HashMap Text Text)

Metadata to assign to the configuration profile. Tags help organize and categorize your AppConfig resources. Each tag consists of a key and an optional value, both of which you define.

applicationId :: Text

The application ID.

name :: Text

A name for the configuration profile.

locationUri :: Text

A URI to locate the configuration. You can specify a Systems Manager (SSM) document, an SSM Parameter Store parameter, or an Amazon S3 object. For an SSM document, specify either the document name in the format ssm-document://<Document_name> or the Amazon Resource Name (ARN). For a parameter, specify either the parameter name in the format ssm-parameter://<Parameter_name> or the ARN. For an Amazon S3 object, specify the URI in the following format: s3://<bucket>/<objectKey> . Here is an example: s3://my-bucket/my-app/us-east-1/my-config.json

newCreateConfigurationProfile

Arguments

:: Text
$sel:applicationId:CreateConfigurationProfile'

-> Text
$sel:name:CreateConfigurationProfile'

-> Text
$sel:locationUri:CreateConfigurationProfile'

-> CreateConfigurationProfile

Create a value of CreateConfigurationProfile with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:retrievalRoleArn:CreateConfigurationProfile', createConfigurationProfile_retrievalRoleArn - The ARN of an IAM role with permission to access the configuration at the specified LocationUri.

$sel:validators:CreateConfigurationProfile', createConfigurationProfile_validators - A list of methods for validating the configuration.

$sel:description:CreateConfigurationProfile', createConfigurationProfile_description - A description of the configuration profile.

$sel:tags:CreateConfigurationProfile', createConfigurationProfile_tags - Metadata to assign to the configuration profile. Tags help organize and categorize your AppConfig resources. Each tag consists of a key and an optional value, both of which you define.

$sel:applicationId:CreateConfigurationProfile', createConfigurationProfile_applicationId - The application ID.

$sel:name:CreateConfigurationProfile', createConfigurationProfile_name - A name for the configuration profile.

$sel:locationUri:CreateConfigurationProfile', createConfigurationProfile_locationUri - A URI to locate the configuration. You can specify a Systems Manager (SSM) document, an SSM Parameter Store parameter, or an Amazon S3 object. For an SSM document, specify either the document name in the format ssm-document://<Document_name> or the Amazon Resource Name (ARN). For a parameter, specify either the parameter name in the format ssm-parameter://<Parameter_name> or the ARN. For an Amazon S3 object, specify the URI in the following format: s3://<bucket>/<objectKey> . Here is an example: s3://my-bucket/my-app/us-east-1/my-config.json

Request Lenses

createConfigurationProfile_retrievalRoleArn :: Lens' CreateConfigurationProfile (Maybe Text)

The ARN of an IAM role with permission to access the configuration at the specified LocationUri.

createConfigurationProfile_validators :: Lens' CreateConfigurationProfile (Maybe [Validator])

A list of methods for validating the configuration.

createConfigurationProfile_description :: Lens' CreateConfigurationProfile (Maybe Text)

A description of the configuration profile.

createConfigurationProfile_tags :: Lens' CreateConfigurationProfile (Maybe (HashMap Text Text))

Metadata to assign to the configuration profile. Tags help organize and categorize your AppConfig resources. Each tag consists of a key and an optional value, both of which you define.

createConfigurationProfile_applicationId :: Lens' CreateConfigurationProfile Text

The application ID.

createConfigurationProfile_name :: Lens' CreateConfigurationProfile Text

A name for the configuration profile.

createConfigurationProfile_locationUri :: Lens' CreateConfigurationProfile Text

A URI to locate the configuration. You can specify a Systems Manager (SSM) document, an SSM Parameter Store parameter, or an Amazon S3 object. For an SSM document, specify either the document name in the format ssm-document://<Document_name> or the Amazon Resource Name (ARN). For a parameter, specify either the parameter name in the format ssm-parameter://<Parameter_name> or the ARN. For an Amazon S3 object, specify the URI in the following format: s3://<bucket>/<objectKey> . Here is an example: s3://my-bucket/my-app/us-east-1/my-config.json

Destructuring the Response

data ConfigurationProfile

See: newConfigurationProfile smart constructor.

Constructors

ConfigurationProfile'

Fields

retrievalRoleArn :: Maybe Text

The ARN of an IAM role with permission to access the configuration at the specified LocationUri.

validators :: Maybe [Validator]

A list of methods for validating the configuration.

locationUri :: Maybe Text

The URI location of the configuration.

applicationId :: Maybe Text

The application ID.

name :: Maybe Text

The name of the configuration profile.

id :: Maybe Text

The configuration profile ID.

description :: Maybe Text

The configuration profile description.

newConfigurationProfile :: ConfigurationProfile

Create a value of ConfigurationProfile with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:retrievalRoleArn:ConfigurationProfile', configurationProfile_retrievalRoleArn - The ARN of an IAM role with permission to access the configuration at the specified LocationUri.

$sel:validators:ConfigurationProfile', configurationProfile_validators - A list of methods for validating the configuration.

$sel:locationUri:ConfigurationProfile', configurationProfile_locationUri - The URI location of the configuration.

$sel:applicationId:ConfigurationProfile', configurationProfile_applicationId - The application ID.

$sel:name:ConfigurationProfile', configurationProfile_name - The name of the configuration profile.

$sel:id:ConfigurationProfile', configurationProfile_id - The configuration profile ID.

$sel:description:ConfigurationProfile', configurationProfile_description - The configuration profile description.

Response Lenses

configurationProfile_retrievalRoleArn :: Lens' ConfigurationProfile (Maybe Text)

The ARN of an IAM role with permission to access the configuration at the specified LocationUri.

configurationProfile_validators :: Lens' ConfigurationProfile (Maybe [Validator])

A list of methods for validating the configuration.

configurationProfile_locationUri :: Lens' ConfigurationProfile (Maybe Text)

The URI location of the configuration.

configurationProfile_applicationId :: Lens' ConfigurationProfile (Maybe Text)

The application ID.

configurationProfile_name :: Lens' ConfigurationProfile (Maybe Text)

The name of the configuration profile.

configurationProfile_id :: Lens' ConfigurationProfile (Maybe Text)

The configuration profile ID.

configurationProfile_description :: Lens' ConfigurationProfile (Maybe Text)

The configuration profile description.