Amazonka.AppConfig.GetApplication

Description

Retrieve information about an application.

Creating a Request

data GetApplication

See: newGetApplication smart constructor.

Constructors

GetApplication'

Fields

applicationId :: Text

The ID of the application you want to get.

newGetApplication

Arguments

:: Text
$sel:applicationId:GetApplication'

-> GetApplication

Create a value of GetApplication with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:applicationId:GetApplication', getApplication_applicationId - The ID of the application you want to get.

Request Lenses

getApplication_applicationId :: Lens' GetApplication Text

The ID of the application you want to get.

Destructuring the Response

data Application

See: newApplication smart constructor.

Constructors

Application'

Fields

name :: Maybe Text

The application name.

id :: Maybe Text

The application ID.

description :: Maybe Text

The description of the application.

newApplication :: Application

Create a value of Application with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:name:Application', application_name - The application name.

$sel:id:Application', application_id - The application ID.

$sel:description:Application', application_description - The description of the application.

Response Lenses

application_name :: Lens' Application (Maybe Text)

The application name.

application_id :: Lens' Application (Maybe Text)

The application ID.

application_description :: Lens' Application (Maybe Text)

The description of the application.