Amazonka.AppConfig.GetHostedConfigurationVersion

Description

Get information about a specific configuration version.

Creating a Request

data GetHostedConfigurationVersion

See: newGetHostedConfigurationVersion smart constructor.

Constructors

GetHostedConfigurationVersion'

Fields

applicationId :: Text

The application ID.

configurationProfileId :: Text

The configuration profile ID.

versionNumber :: Int

The version.

newGetHostedConfigurationVersion

Arguments

:: Text
$sel:applicationId:GetHostedConfigurationVersion'

-> Text
$sel:configurationProfileId:GetHostedConfigurationVersion'

-> Int
$sel:versionNumber:GetHostedConfigurationVersion'

-> GetHostedConfigurationVersion

Create a value of GetHostedConfigurationVersion with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:applicationId:GetHostedConfigurationVersion', getHostedConfigurationVersion_applicationId - The application ID.

$sel:configurationProfileId:GetHostedConfigurationVersion', getHostedConfigurationVersion_configurationProfileId - The configuration profile ID.

$sel:versionNumber:GetHostedConfigurationVersion', getHostedConfigurationVersion_versionNumber - The version.

Request Lenses

getHostedConfigurationVersion_applicationId :: Lens' GetHostedConfigurationVersion Text

The application ID.

getHostedConfigurationVersion_configurationProfileId :: Lens' GetHostedConfigurationVersion Text

The configuration profile ID.

getHostedConfigurationVersion_versionNumber :: Lens' GetHostedConfigurationVersion Int

The version.

Destructuring the Response

data HostedConfigurationVersion

See: newHostedConfigurationVersion smart constructor.

Constructors

HostedConfigurationVersion'

Fields

content :: Maybe (Sensitive ByteString)

The content of the configuration or the configuration data.

versionNumber :: Maybe Int

The configuration version.

applicationId :: Maybe Text

The application ID.

description :: Maybe Text

A description of the configuration.

configurationProfileId :: Maybe Text

The configuration profile ID.

contentType :: Maybe Text

A standard MIME type describing the format of the configuration content. For more information, see Content-Type.

newHostedConfigurationVersion :: HostedConfigurationVersion

Create a value of HostedConfigurationVersion with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:content:HostedConfigurationVersion', hostedConfigurationVersion_content - The content of the configuration or the configuration data.

$sel:versionNumber:HostedConfigurationVersion', hostedConfigurationVersion_versionNumber - The configuration version.

$sel:applicationId:HostedConfigurationVersion', hostedConfigurationVersion_applicationId - The application ID.

$sel:description:HostedConfigurationVersion', hostedConfigurationVersion_description - A description of the configuration.

$sel:configurationProfileId:HostedConfigurationVersion', hostedConfigurationVersion_configurationProfileId - The configuration profile ID.

$sel:contentType:HostedConfigurationVersion', hostedConfigurationVersion_contentType - A standard MIME type describing the format of the configuration content. For more information, see Content-Type.

Response Lenses

hostedConfigurationVersion_content :: Lens' HostedConfigurationVersion (Maybe ByteString)

The content of the configuration or the configuration data.

hostedConfigurationVersion_versionNumber :: Lens' HostedConfigurationVersion (Maybe Int)

The configuration version.

hostedConfigurationVersion_applicationId :: Lens' HostedConfigurationVersion (Maybe Text)

The application ID.

hostedConfigurationVersion_description :: Lens' HostedConfigurationVersion (Maybe Text)

A description of the configuration.

hostedConfigurationVersion_configurationProfileId :: Lens' HostedConfigurationVersion (Maybe Text)

The configuration profile ID.

hostedConfigurationVersion_contentType :: Lens' HostedConfigurationVersion (Maybe Text)

A standard MIME type describing the format of the configuration content. For more information, see Content-Type.