Amazonka.AppConfig.Lens

Description

Operations

ListEnvironments

listEnvironments_nextToken :: Lens' ListEnvironments (Maybe Text)

A token to start the list. Use this token to get the next set of results.

listEnvironments_maxResults :: Lens' ListEnvironments (Maybe Natural)

The maximum number of items to return for this call. The call also returns a token that you can specify in a subsequent call to get the next set of results.

listEnvironments_applicationId :: Lens' ListEnvironments Text

The application ID.

listEnvironmentsResponse_items :: Lens' ListEnvironmentsResponse (Maybe [Environment])

The elements from this collection.

listEnvironmentsResponse_nextToken :: Lens' ListEnvironmentsResponse (Maybe Text)

The token for the next set of items to return. Use this token to get the next set of results.

listEnvironmentsResponse_httpStatus :: Lens' ListEnvironmentsResponse Int

The response's http status code.

UpdateEnvironment

updateEnvironment_monitors :: Lens' UpdateEnvironment (Maybe [Monitor])

Amazon CloudWatch alarms to monitor during the deployment process.

updateEnvironment_name :: Lens' UpdateEnvironment (Maybe Text)

The name of the environment.

updateEnvironment_description :: Lens' UpdateEnvironment (Maybe Text)

A description of the environment.

updateEnvironment_applicationId :: Lens' UpdateEnvironment Text

The application ID.

updateEnvironment_environmentId :: Lens' UpdateEnvironment Text

The environment ID.

environment_state :: Lens' Environment (Maybe EnvironmentState)

The state of the environment. An environment can be in one of the following states: READY_FOR_DEPLOYMENT, DEPLOYING, ROLLING_BACK, or ROLLED_BACK

environment_monitors :: Lens' Environment (Maybe [Monitor])

Amazon CloudWatch alarms monitored during the deployment.

environment_applicationId :: Lens' Environment (Maybe Text)

The application ID.

environment_name :: Lens' Environment (Maybe Text)

The name of the environment.

environment_id :: Lens' Environment (Maybe Text)

The environment ID.

environment_description :: Lens' Environment (Maybe Text)

The description of the environment.

DeleteEnvironment

deleteEnvironment_applicationId :: Lens' DeleteEnvironment Text

The application ID that includes the environment you want to delete.

deleteEnvironment_environmentId :: Lens' DeleteEnvironment Text

The ID of the environment you want to delete.

GetDeploymentStrategy

getDeploymentStrategy_deploymentStrategyId :: Lens' GetDeploymentStrategy Text

The ID of the deployment strategy to get.

deploymentStrategy_growthFactor :: Lens' DeploymentStrategy (Maybe Double)

The percentage of targets that received a deployed configuration during each interval.

deploymentStrategy_replicateTo :: Lens' DeploymentStrategy (Maybe ReplicateTo)

Save the deployment strategy to a Systems Manager (SSM) document.

deploymentStrategy_name :: Lens' DeploymentStrategy (Maybe Text)

The name of the deployment strategy.

deploymentStrategy_id :: Lens' DeploymentStrategy (Maybe Text)

The deployment strategy ID.

deploymentStrategy_deploymentDurationInMinutes :: Lens' DeploymentStrategy (Maybe Natural)

Total amount of time the deployment lasted.

deploymentStrategy_finalBakeTimeInMinutes :: Lens' DeploymentStrategy (Maybe Natural)

The amount of time AppConfig monitored for alarms before considering the deployment to be complete and no longer eligible for automatic roll back.

deploymentStrategy_description :: Lens' DeploymentStrategy (Maybe Text)

The description of the deployment strategy.

deploymentStrategy_growthType :: Lens' DeploymentStrategy (Maybe GrowthType)

The algorithm used to define how percentage grew over time.

CreateConfigurationProfile

createConfigurationProfile_retrievalRoleArn :: Lens' CreateConfigurationProfile (Maybe Text)

The ARN of an IAM role with permission to access the configuration at the specified LocationUri.

createConfigurationProfile_validators :: Lens' CreateConfigurationProfile (Maybe [Validator])

A list of methods for validating the configuration.

createConfigurationProfile_description :: Lens' CreateConfigurationProfile (Maybe Text)

A description of the configuration profile.

createConfigurationProfile_tags :: Lens' CreateConfigurationProfile (Maybe (HashMap Text Text))

Metadata to assign to the configuration profile. Tags help organize and categorize your AppConfig resources. Each tag consists of a key and an optional value, both of which you define.

createConfigurationProfile_applicationId :: Lens' CreateConfigurationProfile Text

The application ID.

createConfigurationProfile_name :: Lens' CreateConfigurationProfile Text

A name for the configuration profile.

createConfigurationProfile_locationUri :: Lens' CreateConfigurationProfile Text

A URI to locate the configuration. You can specify a Systems Manager (SSM) document, an SSM Parameter Store parameter, or an Amazon S3 object. For an SSM document, specify either the document name in the format ssm-document://<Document_name> or the Amazon Resource Name (ARN). For a parameter, specify either the parameter name in the format ssm-parameter://<Parameter_name> or the ARN. For an Amazon S3 object, specify the URI in the following format: s3://<bucket>/<objectKey> . Here is an example: s3://my-bucket/my-app/us-east-1/my-config.json

configurationProfile_retrievalRoleArn :: Lens' ConfigurationProfile (Maybe Text)

The ARN of an IAM role with permission to access the configuration at the specified LocationUri.

configurationProfile_validators :: Lens' ConfigurationProfile (Maybe [Validator])

A list of methods for validating the configuration.

configurationProfile_locationUri :: Lens' ConfigurationProfile (Maybe Text)

The URI location of the configuration.

configurationProfile_applicationId :: Lens' ConfigurationProfile (Maybe Text)

The application ID.

configurationProfile_name :: Lens' ConfigurationProfile (Maybe Text)

The name of the configuration profile.

configurationProfile_id :: Lens' ConfigurationProfile (Maybe Text)

The configuration profile ID.

configurationProfile_description :: Lens' ConfigurationProfile (Maybe Text)

The configuration profile description.

GetDeployment

getDeployment_applicationId :: Lens' GetDeployment Text

The ID of the application that includes the deployment you want to get.

getDeployment_environmentId :: Lens' GetDeployment Text

The ID of the environment that includes the deployment you want to get.

getDeployment_deploymentNumber :: Lens' GetDeployment Int

The sequence number of the deployment.

deployment_growthFactor :: Lens' Deployment (Maybe Double)

The percentage of targets to receive a deployed configuration during each interval.

deployment_configurationName :: Lens' Deployment (Maybe Text)

The name of the configuration.

deployment_state :: Lens' Deployment (Maybe DeploymentState)

The state of the deployment.

deployment_deploymentStrategyId :: Lens' Deployment (Maybe Text)

The ID of the deployment strategy that was deployed.

deployment_deploymentNumber :: Lens' Deployment (Maybe Int)

The sequence number of the deployment.

deployment_configurationVersion :: Lens' Deployment (Maybe Text)

The configuration version that was deployed.

deployment_eventLog :: Lens' Deployment (Maybe [DeploymentEvent])

A list containing all events related to a deployment. The most recent events are displayed first.

deployment_percentageComplete :: Lens' Deployment (Maybe Double)

The percentage of targets for which the deployment is available.

deployment_startedAt :: Lens' Deployment (Maybe UTCTime)

The time the deployment started.

deployment_applicationId :: Lens' Deployment (Maybe Text)

The ID of the application that was deployed.

deployment_deploymentDurationInMinutes :: Lens' Deployment (Maybe Natural)

Total amount of time the deployment lasted.

deployment_environmentId :: Lens' Deployment (Maybe Text)

The ID of the environment that was deployed.

deployment_completedAt :: Lens' Deployment (Maybe UTCTime)

The time the deployment completed.

deployment_configurationLocationUri :: Lens' Deployment (Maybe Text)

Information about the source location of the configuration.

deployment_finalBakeTimeInMinutes :: Lens' Deployment (Maybe Natural)

The amount of time AppConfig monitored for alarms before considering the deployment to be complete and no longer eligible for automatic roll back.

deployment_description :: Lens' Deployment (Maybe Text)

The description of the deployment.

deployment_configurationProfileId :: Lens' Deployment (Maybe Text)

The ID of the configuration profile that was deployed.

deployment_growthType :: Lens' Deployment (Maybe GrowthType)

The algorithm used to define how percentage grew over time.

UpdateConfigurationProfile

updateConfigurationProfile_retrievalRoleArn :: Lens' UpdateConfigurationProfile (Maybe Text)

The ARN of an IAM role with permission to access the configuration at the specified LocationUri.

updateConfigurationProfile_validators :: Lens' UpdateConfigurationProfile (Maybe [Validator])

A list of methods for validating the configuration.

updateConfigurationProfile_name :: Lens' UpdateConfigurationProfile (Maybe Text)

The name of the configuration profile.

updateConfigurationProfile_description :: Lens' UpdateConfigurationProfile (Maybe Text)

A description of the configuration profile.

updateConfigurationProfile_applicationId :: Lens' UpdateConfigurationProfile Text

The application ID.

updateConfigurationProfile_configurationProfileId :: Lens' UpdateConfigurationProfile Text

The ID of the configuration profile.

configurationProfile_retrievalRoleArn :: Lens' ConfigurationProfile (Maybe Text)

The ARN of an IAM role with permission to access the configuration at the specified LocationUri.

configurationProfile_validators :: Lens' ConfigurationProfile (Maybe [Validator])

A list of methods for validating the configuration.

configurationProfile_locationUri :: Lens' ConfigurationProfile (Maybe Text)

The URI location of the configuration.

configurationProfile_applicationId :: Lens' ConfigurationProfile (Maybe Text)

The application ID.

configurationProfile_name :: Lens' ConfigurationProfile (Maybe Text)

The name of the configuration profile.

configurationProfile_id :: Lens' ConfigurationProfile (Maybe Text)

The configuration profile ID.

configurationProfile_description :: Lens' ConfigurationProfile (Maybe Text)

The configuration profile description.

DeleteConfigurationProfile

deleteConfigurationProfile_applicationId :: Lens' DeleteConfigurationProfile Text

The application ID that includes the configuration profile you want to delete.

deleteConfigurationProfile_configurationProfileId :: Lens' DeleteConfigurationProfile Text

The ID of the configuration profile you want to delete.

ListTagsForResource

listTagsForResource_resourceArn :: Lens' ListTagsForResource Text

The resource ARN.

listTagsForResourceResponse_tags :: Lens' ListTagsForResourceResponse (Maybe (HashMap Text Text))

Metadata to assign to AppConfig resources. Tags help organize and categorize your AppConfig resources. Each tag consists of a key and an optional value, both of which you define.

listTagsForResourceResponse_httpStatus :: Lens' ListTagsForResourceResponse Int

The response's http status code.

ListHostedConfigurationVersions

listHostedConfigurationVersions_nextToken :: Lens' ListHostedConfigurationVersions (Maybe Text)

A token to start the list. Use this token to get the next set of results.

listHostedConfigurationVersions_maxResults :: Lens' ListHostedConfigurationVersions (Maybe Natural)

The maximum number of items to return for this call. The call also returns a token that you can specify in a subsequent call to get the next set of results.

listHostedConfigurationVersions_applicationId :: Lens' ListHostedConfigurationVersions Text

The application ID.

listHostedConfigurationVersions_configurationProfileId :: Lens' ListHostedConfigurationVersions Text

The configuration profile ID.

listHostedConfigurationVersionsResponse_items :: Lens' ListHostedConfigurationVersionsResponse (Maybe [HostedConfigurationVersionSummary])

The elements from this collection.

listHostedConfigurationVersionsResponse_nextToken :: Lens' ListHostedConfigurationVersionsResponse (Maybe Text)

The token for the next set of items to return. Use this token to get the next set of results.

listHostedConfigurationVersionsResponse_httpStatus :: Lens' ListHostedConfigurationVersionsResponse Int

The response's http status code.

GetConfigurationProfile

getConfigurationProfile_applicationId :: Lens' GetConfigurationProfile Text

The ID of the application that includes the configuration profile you want to get.

getConfigurationProfile_configurationProfileId :: Lens' GetConfigurationProfile Text

The ID of the configuration profile you want to get.

configurationProfile_retrievalRoleArn :: Lens' ConfigurationProfile (Maybe Text)

The ARN of an IAM role with permission to access the configuration at the specified LocationUri.

configurationProfile_validators :: Lens' ConfigurationProfile (Maybe [Validator])

A list of methods for validating the configuration.

configurationProfile_locationUri :: Lens' ConfigurationProfile (Maybe Text)

The URI location of the configuration.

configurationProfile_applicationId :: Lens' ConfigurationProfile (Maybe Text)

The application ID.

configurationProfile_name :: Lens' ConfigurationProfile (Maybe Text)

The name of the configuration profile.

configurationProfile_id :: Lens' ConfigurationProfile (Maybe Text)

The configuration profile ID.

configurationProfile_description :: Lens' ConfigurationProfile (Maybe Text)

The configuration profile description.

DeleteApplication

deleteApplication_applicationId :: Lens' DeleteApplication Text

The ID of the application to delete.

UpdateApplication

updateApplication_name :: Lens' UpdateApplication (Maybe Text)

The name of the application.

updateApplication_description :: Lens' UpdateApplication (Maybe Text)

A description of the application.

updateApplication_applicationId :: Lens' UpdateApplication Text

The application ID.

application_name :: Lens' Application (Maybe Text)

The application name.

application_id :: Lens' Application (Maybe Text)

The application ID.

application_description :: Lens' Application (Maybe Text)

The description of the application.

UpdateDeploymentStrategy

updateDeploymentStrategy_growthFactor :: Lens' UpdateDeploymentStrategy (Maybe Double)

The percentage of targets to receive a deployed configuration during each interval.

updateDeploymentStrategy_deploymentDurationInMinutes :: Lens' UpdateDeploymentStrategy (Maybe Natural)

Total amount of time for a deployment to last.

updateDeploymentStrategy_finalBakeTimeInMinutes :: Lens' UpdateDeploymentStrategy (Maybe Natural)

The amount of time AppConfig monitors for alarms before considering the deployment to be complete and no longer eligible for automatic roll back.

updateDeploymentStrategy_description :: Lens' UpdateDeploymentStrategy (Maybe Text)

A description of the deployment strategy.

updateDeploymentStrategy_growthType :: Lens' UpdateDeploymentStrategy (Maybe GrowthType)

The algorithm used to define how percentage grows over time. AWS AppConfig supports the following growth types:

Linear: For this type, AppConfig processes the deployment by increments of the growth factor evenly distributed over the deployment time. For example, a linear deployment that uses a growth factor of 20 initially makes the configuration available to 20 percent of the targets. After 1/5th of the deployment time has passed, the system updates the percentage to 40 percent. This continues until 100% of the targets are set to receive the deployed configuration.

Exponential: For this type, AppConfig processes the deployment exponentially using the following formula: G*(2^N). In this formula, G is the growth factor specified by the user and N is the number of steps until the configuration is deployed to all targets. For example, if you specify a growth factor of 2, then the system rolls out the configuration as follows:

2*(2^0)

2*(2^1)

2*(2^2)

Expressed numerically, the deployment rolls out as follows: 2% of the targets, 4% of the targets, 8% of the targets, and continues until the configuration has been deployed to all targets.

updateDeploymentStrategy_deploymentStrategyId :: Lens' UpdateDeploymentStrategy Text

The deployment strategy ID.

deploymentStrategy_growthFactor :: Lens' DeploymentStrategy (Maybe Double)

The percentage of targets that received a deployed configuration during each interval.

deploymentStrategy_replicateTo :: Lens' DeploymentStrategy (Maybe ReplicateTo)

Save the deployment strategy to a Systems Manager (SSM) document.

deploymentStrategy_name :: Lens' DeploymentStrategy (Maybe Text)

The name of the deployment strategy.

deploymentStrategy_id :: Lens' DeploymentStrategy (Maybe Text)

The deployment strategy ID.

deploymentStrategy_deploymentDurationInMinutes :: Lens' DeploymentStrategy (Maybe Natural)

Total amount of time the deployment lasted.

deploymentStrategy_finalBakeTimeInMinutes :: Lens' DeploymentStrategy (Maybe Natural)

The amount of time AppConfig monitored for alarms before considering the deployment to be complete and no longer eligible for automatic roll back.

deploymentStrategy_description :: Lens' DeploymentStrategy (Maybe Text)

The description of the deployment strategy.

deploymentStrategy_growthType :: Lens' DeploymentStrategy (Maybe GrowthType)

The algorithm used to define how percentage grew over time.

DeleteDeploymentStrategy

deleteDeploymentStrategy_deploymentStrategyId :: Lens' DeleteDeploymentStrategy Text

The ID of the deployment strategy you want to delete.

CreateApplication

createApplication_description :: Lens' CreateApplication (Maybe Text)

A description of the application.

createApplication_tags :: Lens' CreateApplication (Maybe (HashMap Text Text))

Metadata to assign to the application. Tags help organize and categorize your AppConfig resources. Each tag consists of a key and an optional value, both of which you define.

createApplication_name :: Lens' CreateApplication Text

A name for the application.

application_name :: Lens' Application (Maybe Text)

The application name.

application_id :: Lens' Application (Maybe Text)

The application ID.

application_description :: Lens' Application (Maybe Text)

The description of the application.

ValidateConfiguration

validateConfiguration_applicationId :: Lens' ValidateConfiguration Text

The application ID.

validateConfiguration_configurationProfileId :: Lens' ValidateConfiguration Text

The configuration profile ID.

validateConfiguration_configurationVersion :: Lens' ValidateConfiguration Text

The version of the configuration to validate.

StopDeployment

stopDeployment_applicationId :: Lens' StopDeployment Text

The application ID.

stopDeployment_environmentId :: Lens' StopDeployment Text

The environment ID.

stopDeployment_deploymentNumber :: Lens' StopDeployment Int

The sequence number of the deployment.

deployment_growthFactor :: Lens' Deployment (Maybe Double)

The percentage of targets to receive a deployed configuration during each interval.

deployment_configurationName :: Lens' Deployment (Maybe Text)

The name of the configuration.

deployment_state :: Lens' Deployment (Maybe DeploymentState)

The state of the deployment.

deployment_deploymentStrategyId :: Lens' Deployment (Maybe Text)

The ID of the deployment strategy that was deployed.

deployment_deploymentNumber :: Lens' Deployment (Maybe Int)

The sequence number of the deployment.

deployment_configurationVersion :: Lens' Deployment (Maybe Text)

The configuration version that was deployed.

deployment_eventLog :: Lens' Deployment (Maybe [DeploymentEvent])

A list containing all events related to a deployment. The most recent events are displayed first.

deployment_percentageComplete :: Lens' Deployment (Maybe Double)

The percentage of targets for which the deployment is available.

deployment_startedAt :: Lens' Deployment (Maybe UTCTime)

The time the deployment started.

deployment_applicationId :: Lens' Deployment (Maybe Text)

The ID of the application that was deployed.

deployment_deploymentDurationInMinutes :: Lens' Deployment (Maybe Natural)

Total amount of time the deployment lasted.

deployment_environmentId :: Lens' Deployment (Maybe Text)

The ID of the environment that was deployed.

deployment_completedAt :: Lens' Deployment (Maybe UTCTime)

The time the deployment completed.

deployment_configurationLocationUri :: Lens' Deployment (Maybe Text)

Information about the source location of the configuration.

deployment_finalBakeTimeInMinutes :: Lens' Deployment (Maybe Natural)

The amount of time AppConfig monitored for alarms before considering the deployment to be complete and no longer eligible for automatic roll back.

deployment_description :: Lens' Deployment (Maybe Text)

The description of the deployment.

deployment_configurationProfileId :: Lens' Deployment (Maybe Text)

The ID of the configuration profile that was deployed.

deployment_growthType :: Lens' Deployment (Maybe GrowthType)

The algorithm used to define how percentage grew over time.

GetApplication

getApplication_applicationId :: Lens' GetApplication Text

The ID of the application you want to get.

application_name :: Lens' Application (Maybe Text)

The application name.

application_id :: Lens' Application (Maybe Text)

The application ID.

application_description :: Lens' Application (Maybe Text)

The description of the application.

CreateHostedConfigurationVersion

createHostedConfigurationVersion_latestVersionNumber :: Lens' CreateHostedConfigurationVersion (Maybe Int)

An optional locking token used to prevent race conditions from overwriting configuration updates when creating a new version. To ensure your data is not overwritten when creating multiple hosted configuration versions in rapid succession, specify the version of the latest hosted configuration version.

createHostedConfigurationVersion_description :: Lens' CreateHostedConfigurationVersion (Maybe Text)

A description of the configuration.

createHostedConfigurationVersion_applicationId :: Lens' CreateHostedConfigurationVersion Text

The application ID.

createHostedConfigurationVersion_configurationProfileId :: Lens' CreateHostedConfigurationVersion Text

The configuration profile ID.

createHostedConfigurationVersion_content :: Lens' CreateHostedConfigurationVersion ByteString

The content of the configuration or the configuration data.

createHostedConfigurationVersion_contentType :: Lens' CreateHostedConfigurationVersion Text

A standard MIME type describing the format of the configuration content. For more information, see Content-Type.

hostedConfigurationVersion_content :: Lens' HostedConfigurationVersion (Maybe ByteString)

The content of the configuration or the configuration data.

hostedConfigurationVersion_versionNumber :: Lens' HostedConfigurationVersion (Maybe Int)

The configuration version.

hostedConfigurationVersion_applicationId :: Lens' HostedConfigurationVersion (Maybe Text)

The application ID.

hostedConfigurationVersion_description :: Lens' HostedConfigurationVersion (Maybe Text)

A description of the configuration.

hostedConfigurationVersion_configurationProfileId :: Lens' HostedConfigurationVersion (Maybe Text)

The configuration profile ID.

hostedConfigurationVersion_contentType :: Lens' HostedConfigurationVersion (Maybe Text)

A standard MIME type describing the format of the configuration content. For more information, see Content-Type.

ListConfigurationProfiles

listConfigurationProfiles_nextToken :: Lens' ListConfigurationProfiles (Maybe Text)

A token to start the list. Use this token to get the next set of results.

listConfigurationProfiles_maxResults :: Lens' ListConfigurationProfiles (Maybe Natural)

The maximum number of items to return for this call. The call also returns a token that you can specify in a subsequent call to get the next set of results.

listConfigurationProfiles_applicationId :: Lens' ListConfigurationProfiles Text

The application ID.

listConfigurationProfilesResponse_items :: Lens' ListConfigurationProfilesResponse (Maybe [ConfigurationProfileSummary])

The elements from this collection.

listConfigurationProfilesResponse_nextToken :: Lens' ListConfigurationProfilesResponse (Maybe Text)

The token for the next set of items to return. Use this token to get the next set of results.

listConfigurationProfilesResponse_httpStatus :: Lens' ListConfigurationProfilesResponse Int

The response's http status code.

DeleteHostedConfigurationVersion

deleteHostedConfigurationVersion_applicationId :: Lens' DeleteHostedConfigurationVersion Text

The application ID.

deleteHostedConfigurationVersion_configurationProfileId :: Lens' DeleteHostedConfigurationVersion Text

The configuration profile ID.

deleteHostedConfigurationVersion_versionNumber :: Lens' DeleteHostedConfigurationVersion Int

The versions number to delete.

GetHostedConfigurationVersion

getHostedConfigurationVersion_applicationId :: Lens' GetHostedConfigurationVersion Text

The application ID.

getHostedConfigurationVersion_configurationProfileId :: Lens' GetHostedConfigurationVersion Text

The configuration profile ID.

getHostedConfigurationVersion_versionNumber :: Lens' GetHostedConfigurationVersion Int

The version.

hostedConfigurationVersion_content :: Lens' HostedConfigurationVersion (Maybe ByteString)

The content of the configuration or the configuration data.

hostedConfigurationVersion_versionNumber :: Lens' HostedConfigurationVersion (Maybe Int)

The configuration version.

hostedConfigurationVersion_applicationId :: Lens' HostedConfigurationVersion (Maybe Text)

The application ID.

hostedConfigurationVersion_description :: Lens' HostedConfigurationVersion (Maybe Text)

A description of the configuration.

hostedConfigurationVersion_configurationProfileId :: Lens' HostedConfigurationVersion (Maybe Text)

The configuration profile ID.

hostedConfigurationVersion_contentType :: Lens' HostedConfigurationVersion (Maybe Text)

A standard MIME type describing the format of the configuration content. For more information, see Content-Type.

ListDeployments

listDeployments_nextToken :: Lens' ListDeployments (Maybe Text)

A token to start the list. Use this token to get the next set of results.

listDeployments_maxResults :: Lens' ListDeployments (Maybe Natural)

The maximum number of items to return for this call. The call also returns a token that you can specify in a subsequent call to get the next set of results.

listDeployments_applicationId :: Lens' ListDeployments Text

The application ID.

listDeployments_environmentId :: Lens' ListDeployments Text

The environment ID.

listDeploymentsResponse_items :: Lens' ListDeploymentsResponse (Maybe [DeploymentSummary])

The elements from this collection.

listDeploymentsResponse_nextToken :: Lens' ListDeploymentsResponse (Maybe Text)

The token for the next set of items to return. Use this token to get the next set of results.

listDeploymentsResponse_httpStatus :: Lens' ListDeploymentsResponse Int

The response's http status code.

TagResource

tagResource_resourceArn :: Lens' TagResource Text

The ARN of the resource for which to retrieve tags.

tagResource_tags :: Lens' TagResource (HashMap Text Text)

The key-value string map. The valid character set is [a-zA-Z+-=._:/]. The tag key can be up to 128 characters and must not start with aws:. The tag value can be up to 256 characters.

GetEnvironment

getEnvironment_applicationId :: Lens' GetEnvironment Text

The ID of the application that includes the environment you want to get.

getEnvironment_environmentId :: Lens' GetEnvironment Text

The ID of the environment you wnat to get.

environment_state :: Lens' Environment (Maybe EnvironmentState)

The state of the environment. An environment can be in one of the following states: READY_FOR_DEPLOYMENT, DEPLOYING, ROLLING_BACK, or ROLLED_BACK

environment_monitors :: Lens' Environment (Maybe [Monitor])

Amazon CloudWatch alarms monitored during the deployment.

environment_applicationId :: Lens' Environment (Maybe Text)

The application ID.

environment_name :: Lens' Environment (Maybe Text)

The name of the environment.

environment_id :: Lens' Environment (Maybe Text)

The environment ID.

environment_description :: Lens' Environment (Maybe Text)

The description of the environment.

ListApplications

listApplications_nextToken :: Lens' ListApplications (Maybe Text)

A token to start the list. Use this token to get the next set of results.

listApplications_maxResults :: Lens' ListApplications (Maybe Natural)

The maximum number of items to return for this call. The call also returns a token that you can specify in a subsequent call to get the next set of results.

listApplicationsResponse_items :: Lens' ListApplicationsResponse (Maybe [Application])

The elements from this collection.

listApplicationsResponse_nextToken :: Lens' ListApplicationsResponse (Maybe Text)

The token for the next set of items to return. Use this token to get the next set of results.

listApplicationsResponse_httpStatus :: Lens' ListApplicationsResponse Int

The response's http status code.

UntagResource

untagResource_resourceArn :: Lens' UntagResource Text

The ARN of the resource for which to remove tags.

untagResource_tagKeys :: Lens' UntagResource [Text]

The tag keys to delete.

ListDeploymentStrategies

listDeploymentStrategies_nextToken :: Lens' ListDeploymentStrategies (Maybe Text)

A token to start the list. Use this token to get the next set of results.

listDeploymentStrategies_maxResults :: Lens' ListDeploymentStrategies (Maybe Natural)

The maximum number of items to return for this call. The call also returns a token that you can specify in a subsequent call to get the next set of results.

listDeploymentStrategiesResponse_items :: Lens' ListDeploymentStrategiesResponse (Maybe [DeploymentStrategy])

The elements from this collection.

listDeploymentStrategiesResponse_nextToken :: Lens' ListDeploymentStrategiesResponse (Maybe Text)

The token for the next set of items to return. Use this token to get the next set of results.

listDeploymentStrategiesResponse_httpStatus :: Lens' ListDeploymentStrategiesResponse Int

The response's http status code.

GetConfiguration

getConfiguration_clientConfigurationVersion :: Lens' GetConfiguration (Maybe Text)

The configuration version returned in the most recent GetConfiguration response.

AWS AppConfig uses the value of the ClientConfigurationVersion parameter to identify the configuration version on your clients. If you don’t send ClientConfigurationVersion with each call to GetConfiguration, your clients receive the current configuration. You are charged each time your clients receive a configuration.

To avoid excess charges, we recommend that you include the ClientConfigurationVersion value with every call to GetConfiguration. This value must be saved on your client. Subsequent calls to GetConfiguration must pass this value by using the ClientConfigurationVersion parameter.

For more information about working with configurations, see Retrieving the Configuration in the AWS AppConfig User Guide.

getConfiguration_application :: Lens' GetConfiguration Text

The application to get. Specify either the application name or the application ID.

getConfiguration_environment :: Lens' GetConfiguration Text

The environment to get. Specify either the environment name or the environment ID.

getConfiguration_configuration :: Lens' GetConfiguration Text

The configuration to get. Specify either the configuration name or the configuration ID.

getConfiguration_clientId :: Lens' GetConfiguration Text

A unique ID to identify the client for the configuration. This ID enables AppConfig to deploy the configuration in intervals, as defined in the deployment strategy.

getConfigurationResponse_configurationVersion :: Lens' GetConfigurationResponse (Maybe Text)

The configuration version.

getConfigurationResponse_content :: Lens' GetConfigurationResponse (Maybe ByteString)

The content of the configuration or the configuration data.

getConfigurationResponse_contentType :: Lens' GetConfigurationResponse (Maybe Text)

A standard MIME type describing the format of the configuration content. For more information, see Content-Type.

getConfigurationResponse_httpStatus :: Lens' GetConfigurationResponse Int

The response's http status code.

CreateDeploymentStrategy

createDeploymentStrategy_finalBakeTimeInMinutes :: Lens' CreateDeploymentStrategy (Maybe Natural)

The amount of time AppConfig monitors for alarms before considering the deployment to be complete and no longer eligible for automatic roll back.

createDeploymentStrategy_description :: Lens' CreateDeploymentStrategy (Maybe Text)

A description of the deployment strategy.

createDeploymentStrategy_growthType :: Lens' CreateDeploymentStrategy (Maybe GrowthType)

The algorithm used to define how percentage grows over time. AWS AppConfig supports the following growth types:

Linear: For this type, AppConfig processes the deployment by dividing the total number of targets by the value specified for Step percentage. For example, a linear deployment that uses a Step percentage of 10 deploys the configuration to 10 percent of the hosts. After those deployments are complete, the system deploys the configuration to the next 10 percent. This continues until 100% of the targets have successfully received the configuration.

Exponential: For this type, AppConfig processes the deployment exponentially using the following formula: G*(2^N). In this formula, G is the growth factor specified by the user and N is the number of steps until the configuration is deployed to all targets. For example, if you specify a growth factor of 2, then the system rolls out the configuration as follows:

2*(2^0)

2*(2^1)

2*(2^2)

Expressed numerically, the deployment rolls out as follows: 2% of the targets, 4% of the targets, 8% of the targets, and continues until the configuration has been deployed to all targets.

createDeploymentStrategy_tags :: Lens' CreateDeploymentStrategy (Maybe (HashMap Text Text))

Metadata to assign to the deployment strategy. Tags help organize and categorize your AppConfig resources. Each tag consists of a key and an optional value, both of which you define.

createDeploymentStrategy_name :: Lens' CreateDeploymentStrategy Text

A name for the deployment strategy.

createDeploymentStrategy_deploymentDurationInMinutes :: Lens' CreateDeploymentStrategy Natural

Total amount of time for a deployment to last.

createDeploymentStrategy_growthFactor :: Lens' CreateDeploymentStrategy Double

The percentage of targets to receive a deployed configuration during each interval.

createDeploymentStrategy_replicateTo :: Lens' CreateDeploymentStrategy ReplicateTo

Save the deployment strategy to a Systems Manager (SSM) document.

deploymentStrategy_growthFactor :: Lens' DeploymentStrategy (Maybe Double)

The percentage of targets that received a deployed configuration during each interval.

deploymentStrategy_replicateTo :: Lens' DeploymentStrategy (Maybe ReplicateTo)

Save the deployment strategy to a Systems Manager (SSM) document.

deploymentStrategy_name :: Lens' DeploymentStrategy (Maybe Text)

The name of the deployment strategy.

deploymentStrategy_id :: Lens' DeploymentStrategy (Maybe Text)

The deployment strategy ID.

deploymentStrategy_deploymentDurationInMinutes :: Lens' DeploymentStrategy (Maybe Natural)

Total amount of time the deployment lasted.

deploymentStrategy_finalBakeTimeInMinutes :: Lens' DeploymentStrategy (Maybe Natural)

The amount of time AppConfig monitored for alarms before considering the deployment to be complete and no longer eligible for automatic roll back.

deploymentStrategy_description :: Lens' DeploymentStrategy (Maybe Text)

The description of the deployment strategy.

deploymentStrategy_growthType :: Lens' DeploymentStrategy (Maybe GrowthType)

The algorithm used to define how percentage grew over time.

StartDeployment

startDeployment_description :: Lens' StartDeployment (Maybe Text)

A description of the deployment.

startDeployment_tags :: Lens' StartDeployment (Maybe (HashMap Text Text))

Metadata to assign to the deployment. Tags help organize and categorize your AppConfig resources. Each tag consists of a key and an optional value, both of which you define.

startDeployment_applicationId :: Lens' StartDeployment Text

The application ID.

startDeployment_environmentId :: Lens' StartDeployment Text

The environment ID.

startDeployment_deploymentStrategyId :: Lens' StartDeployment Text

The deployment strategy ID.

startDeployment_configurationProfileId :: Lens' StartDeployment Text

The configuration profile ID.

startDeployment_configurationVersion :: Lens' StartDeployment Text

The configuration version to deploy.

deployment_growthFactor :: Lens' Deployment (Maybe Double)

The percentage of targets to receive a deployed configuration during each interval.

deployment_configurationName :: Lens' Deployment (Maybe Text)

The name of the configuration.

deployment_state :: Lens' Deployment (Maybe DeploymentState)

The state of the deployment.

deployment_deploymentStrategyId :: Lens' Deployment (Maybe Text)

The ID of the deployment strategy that was deployed.

deployment_deploymentNumber :: Lens' Deployment (Maybe Int)

The sequence number of the deployment.

deployment_configurationVersion :: Lens' Deployment (Maybe Text)

The configuration version that was deployed.

deployment_eventLog :: Lens' Deployment (Maybe [DeploymentEvent])

A list containing all events related to a deployment. The most recent events are displayed first.

deployment_percentageComplete :: Lens' Deployment (Maybe Double)

The percentage of targets for which the deployment is available.

deployment_startedAt :: Lens' Deployment (Maybe UTCTime)

The time the deployment started.

deployment_applicationId :: Lens' Deployment (Maybe Text)

The ID of the application that was deployed.

deployment_deploymentDurationInMinutes :: Lens' Deployment (Maybe Natural)

Total amount of time the deployment lasted.

deployment_environmentId :: Lens' Deployment (Maybe Text)

The ID of the environment that was deployed.

deployment_completedAt :: Lens' Deployment (Maybe UTCTime)

The time the deployment completed.

deployment_configurationLocationUri :: Lens' Deployment (Maybe Text)

Information about the source location of the configuration.

deployment_finalBakeTimeInMinutes :: Lens' Deployment (Maybe Natural)

The amount of time AppConfig monitored for alarms before considering the deployment to be complete and no longer eligible for automatic roll back.

deployment_description :: Lens' Deployment (Maybe Text)

The description of the deployment.

deployment_configurationProfileId :: Lens' Deployment (Maybe Text)

The ID of the configuration profile that was deployed.

deployment_growthType :: Lens' Deployment (Maybe GrowthType)

The algorithm used to define how percentage grew over time.

CreateEnvironment

createEnvironment_monitors :: Lens' CreateEnvironment (Maybe [Monitor])

Amazon CloudWatch alarms to monitor during the deployment process.

createEnvironment_description :: Lens' CreateEnvironment (Maybe Text)

A description of the environment.

createEnvironment_tags :: Lens' CreateEnvironment (Maybe (HashMap Text Text))

Metadata to assign to the environment. Tags help organize and categorize your AppConfig resources. Each tag consists of a key and an optional value, both of which you define.

createEnvironment_applicationId :: Lens' CreateEnvironment Text

The application ID.

createEnvironment_name :: Lens' CreateEnvironment Text

A name for the environment.

environment_state :: Lens' Environment (Maybe EnvironmentState)

The state of the environment. An environment can be in one of the following states: READY_FOR_DEPLOYMENT, DEPLOYING, ROLLING_BACK, or ROLLED_BACK

environment_monitors :: Lens' Environment (Maybe [Monitor])

Amazon CloudWatch alarms monitored during the deployment.

environment_applicationId :: Lens' Environment (Maybe Text)

The application ID.

environment_name :: Lens' Environment (Maybe Text)

The name of the environment.

environment_id :: Lens' Environment (Maybe Text)

The environment ID.

environment_description :: Lens' Environment (Maybe Text)

The description of the environment.

Types

Application

application_name :: Lens' Application (Maybe Text)

The application name.

application_id :: Lens' Application (Maybe Text)

The application ID.

application_description :: Lens' Application (Maybe Text)

The description of the application.

ConfigurationProfile

configurationProfile_retrievalRoleArn :: Lens' ConfigurationProfile (Maybe Text)

The ARN of an IAM role with permission to access the configuration at the specified LocationUri.

configurationProfile_validators :: Lens' ConfigurationProfile (Maybe [Validator])

A list of methods for validating the configuration.

configurationProfile_locationUri :: Lens' ConfigurationProfile (Maybe Text)

The URI location of the configuration.

configurationProfile_applicationId :: Lens' ConfigurationProfile (Maybe Text)

The application ID.

configurationProfile_name :: Lens' ConfigurationProfile (Maybe Text)

The name of the configuration profile.

configurationProfile_id :: Lens' ConfigurationProfile (Maybe Text)

The configuration profile ID.

configurationProfile_description :: Lens' ConfigurationProfile (Maybe Text)

The configuration profile description.

ConfigurationProfileSummary

configurationProfileSummary_locationUri :: Lens' ConfigurationProfileSummary (Maybe Text)

The URI location of the configuration.

configurationProfileSummary_applicationId :: Lens' ConfigurationProfileSummary (Maybe Text)

The application ID.

configurationProfileSummary_name :: Lens' ConfigurationProfileSummary (Maybe Text)

The name of the configuration profile.

configurationProfileSummary_id :: Lens' ConfigurationProfileSummary (Maybe Text)

The ID of the configuration profile.

configurationProfileSummary_validatorTypes :: Lens' ConfigurationProfileSummary (Maybe [ValidatorType])

The types of validators in the configuration profile.

Deployment

deployment_growthFactor :: Lens' Deployment (Maybe Double)

The percentage of targets to receive a deployed configuration during each interval.

deployment_configurationName :: Lens' Deployment (Maybe Text)

The name of the configuration.

deployment_state :: Lens' Deployment (Maybe DeploymentState)

The state of the deployment.

deployment_deploymentStrategyId :: Lens' Deployment (Maybe Text)

The ID of the deployment strategy that was deployed.

deployment_deploymentNumber :: Lens' Deployment (Maybe Int)

The sequence number of the deployment.

deployment_configurationVersion :: Lens' Deployment (Maybe Text)

The configuration version that was deployed.

deployment_eventLog :: Lens' Deployment (Maybe [DeploymentEvent])

A list containing all events related to a deployment. The most recent events are displayed first.

deployment_percentageComplete :: Lens' Deployment (Maybe Double)

The percentage of targets for which the deployment is available.

deployment_startedAt :: Lens' Deployment (Maybe UTCTime)

The time the deployment started.

deployment_applicationId :: Lens' Deployment (Maybe Text)

The ID of the application that was deployed.

deployment_deploymentDurationInMinutes :: Lens' Deployment (Maybe Natural)

Total amount of time the deployment lasted.

deployment_environmentId :: Lens' Deployment (Maybe Text)

The ID of the environment that was deployed.

deployment_completedAt :: Lens' Deployment (Maybe UTCTime)

The time the deployment completed.

deployment_configurationLocationUri :: Lens' Deployment (Maybe Text)

Information about the source location of the configuration.

deployment_finalBakeTimeInMinutes :: Lens' Deployment (Maybe Natural)

The amount of time AppConfig monitored for alarms before considering the deployment to be complete and no longer eligible for automatic roll back.

deployment_description :: Lens' Deployment (Maybe Text)

The description of the deployment.

deployment_configurationProfileId :: Lens' Deployment (Maybe Text)

The ID of the configuration profile that was deployed.

deployment_growthType :: Lens' Deployment (Maybe GrowthType)

The algorithm used to define how percentage grew over time.

DeploymentEvent

deploymentEvent_triggeredBy :: Lens' DeploymentEvent (Maybe TriggeredBy)

The entity that triggered the deployment event. Events can be triggered by a user, AWS AppConfig, an Amazon CloudWatch alarm, or an internal error.

deploymentEvent_occurredAt :: Lens' DeploymentEvent (Maybe UTCTime)

The date and time the event occurred.

deploymentEvent_eventType :: Lens' DeploymentEvent (Maybe DeploymentEventType)

The type of deployment event. Deployment event types include the start, stop, or completion of a deployment; a percentage update; the start or stop of a bake period; the start or completion of a rollback.

deploymentEvent_description :: Lens' DeploymentEvent (Maybe Text)

A description of the deployment event. Descriptions include, but are not limited to, the user account or the CloudWatch alarm ARN that initiated a rollback, the percentage of hosts that received the deployment, or in the case of an internal error, a recommendation to attempt a new deployment.

DeploymentStrategy

deploymentStrategy_growthFactor :: Lens' DeploymentStrategy (Maybe Double)

The percentage of targets that received a deployed configuration during each interval.

deploymentStrategy_replicateTo :: Lens' DeploymentStrategy (Maybe ReplicateTo)

Save the deployment strategy to a Systems Manager (SSM) document.

deploymentStrategy_name :: Lens' DeploymentStrategy (Maybe Text)

The name of the deployment strategy.

deploymentStrategy_id :: Lens' DeploymentStrategy (Maybe Text)

The deployment strategy ID.

deploymentStrategy_deploymentDurationInMinutes :: Lens' DeploymentStrategy (Maybe Natural)

Total amount of time the deployment lasted.

deploymentStrategy_finalBakeTimeInMinutes :: Lens' DeploymentStrategy (Maybe Natural)

The amount of time AppConfig monitored for alarms before considering the deployment to be complete and no longer eligible for automatic roll back.

deploymentStrategy_description :: Lens' DeploymentStrategy (Maybe Text)

The description of the deployment strategy.

deploymentStrategy_growthType :: Lens' DeploymentStrategy (Maybe GrowthType)

The algorithm used to define how percentage grew over time.

DeploymentSummary

deploymentSummary_growthFactor :: Lens' DeploymentSummary (Maybe Double)

The percentage of targets to receive a deployed configuration during each interval.

deploymentSummary_configurationName :: Lens' DeploymentSummary (Maybe Text)

The name of the configuration.

deploymentSummary_state :: Lens' DeploymentSummary (Maybe DeploymentState)

The state of the deployment.

deploymentSummary_deploymentNumber :: Lens' DeploymentSummary (Maybe Int)

The sequence number of the deployment.

deploymentSummary_configurationVersion :: Lens' DeploymentSummary (Maybe Text)

The version of the configuration.

deploymentSummary_percentageComplete :: Lens' DeploymentSummary (Maybe Double)

The percentage of targets for which the deployment is available.

deploymentSummary_startedAt :: Lens' DeploymentSummary (Maybe UTCTime)

Time the deployment started.

deploymentSummary_deploymentDurationInMinutes :: Lens' DeploymentSummary (Maybe Natural)

Total amount of time the deployment lasted.

deploymentSummary_completedAt :: Lens' DeploymentSummary (Maybe UTCTime)

Time the deployment completed.

deploymentSummary_finalBakeTimeInMinutes :: Lens' DeploymentSummary (Maybe Natural)

The amount of time AppConfig monitors for alarms before considering the deployment to be complete and no longer eligible for automatic roll back.

deploymentSummary_growthType :: Lens' DeploymentSummary (Maybe GrowthType)

The algorithm used to define how percentage grows over time.

Environment

environment_state :: Lens' Environment (Maybe EnvironmentState)

The state of the environment. An environment can be in one of the following states: READY_FOR_DEPLOYMENT, DEPLOYING, ROLLING_BACK, or ROLLED_BACK

environment_monitors :: Lens' Environment (Maybe [Monitor])

Amazon CloudWatch alarms monitored during the deployment.

environment_applicationId :: Lens' Environment (Maybe Text)

The application ID.

environment_name :: Lens' Environment (Maybe Text)

The name of the environment.

environment_id :: Lens' Environment (Maybe Text)

The environment ID.

environment_description :: Lens' Environment (Maybe Text)

The description of the environment.

HostedConfigurationVersion

hostedConfigurationVersion_content :: Lens' HostedConfigurationVersion (Maybe ByteString)

The content of the configuration or the configuration data.

hostedConfigurationVersion_versionNumber :: Lens' HostedConfigurationVersion (Maybe Int)

The configuration version.

hostedConfigurationVersion_applicationId :: Lens' HostedConfigurationVersion (Maybe Text)

The application ID.

hostedConfigurationVersion_description :: Lens' HostedConfigurationVersion (Maybe Text)

A description of the configuration.

hostedConfigurationVersion_configurationProfileId :: Lens' HostedConfigurationVersion (Maybe Text)

The configuration profile ID.

hostedConfigurationVersion_contentType :: Lens' HostedConfigurationVersion (Maybe Text)

A standard MIME type describing the format of the configuration content. For more information, see Content-Type.

HostedConfigurationVersionSummary

hostedConfigurationVersionSummary_versionNumber :: Lens' HostedConfigurationVersionSummary (Maybe Int)

The configuration version.

hostedConfigurationVersionSummary_applicationId :: Lens' HostedConfigurationVersionSummary (Maybe Text)

The application ID.

hostedConfigurationVersionSummary_description :: Lens' HostedConfigurationVersionSummary (Maybe Text)

A description of the configuration.

hostedConfigurationVersionSummary_configurationProfileId :: Lens' HostedConfigurationVersionSummary (Maybe Text)

The configuration profile ID.

hostedConfigurationVersionSummary_contentType :: Lens' HostedConfigurationVersionSummary (Maybe Text)

A standard MIME type describing the format of the configuration content. For more information, see Content-Type.

Monitor

monitor_alarmRoleArn :: Lens' Monitor (Maybe Text)

ARN of an IAM role for AppConfig to monitor AlarmArn.

monitor_alarmArn :: Lens' Monitor (Maybe Text)

ARN of the Amazon CloudWatch alarm.

Validator

validator_type :: Lens' Validator ValidatorType

AppConfig supports validators of type JSON_SCHEMA and LAMBDA

validator_content :: Lens' Validator Text

Either the JSON Schema content or the Amazon Resource Name (ARN) of an AWS Lambda function.