Amazonka.AppConfig.ListConfigurationProfiles

Description

Lists the configuration profiles for an application.

Creating a Request

data ListConfigurationProfiles

See: newListConfigurationProfiles smart constructor.

Constructors

ListConfigurationProfiles'

Fields

nextToken :: Maybe Text

A token to start the list. Use this token to get the next set of results.

maxResults :: Maybe Natural

The maximum number of items to return for this call. The call also returns a token that you can specify in a subsequent call to get the next set of results.

applicationId :: Text

The application ID.

newListConfigurationProfiles

Arguments

:: Text
$sel:applicationId:ListConfigurationProfiles'

-> ListConfigurationProfiles

Create a value of ListConfigurationProfiles with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:nextToken:ListConfigurationProfiles', listConfigurationProfiles_nextToken - A token to start the list. Use this token to get the next set of results.

$sel:maxResults:ListConfigurationProfiles', listConfigurationProfiles_maxResults - The maximum number of items to return for this call. The call also returns a token that you can specify in a subsequent call to get the next set of results.

$sel:applicationId:ListConfigurationProfiles', listConfigurationProfiles_applicationId - The application ID.

Request Lenses

listConfigurationProfiles_nextToken :: Lens' ListConfigurationProfiles (Maybe Text)

A token to start the list. Use this token to get the next set of results.

listConfigurationProfiles_maxResults :: Lens' ListConfigurationProfiles (Maybe Natural)

The maximum number of items to return for this call. The call also returns a token that you can specify in a subsequent call to get the next set of results.

listConfigurationProfiles_applicationId :: Lens' ListConfigurationProfiles Text

The application ID.

Destructuring the Response

data ListConfigurationProfilesResponse

See: newListConfigurationProfilesResponse smart constructor.

Constructors

ListConfigurationProfilesResponse'

Fields

items :: Maybe [ConfigurationProfileSummary]

The elements from this collection.

nextToken :: Maybe Text

The token for the next set of items to return. Use this token to get the next set of results.

httpStatus :: Int

The response's http status code.

newListConfigurationProfilesResponse

Arguments

:: Int
$sel:httpStatus:ListConfigurationProfilesResponse'

-> ListConfigurationProfilesResponse

Create a value of ListConfigurationProfilesResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:items:ListConfigurationProfilesResponse', listConfigurationProfilesResponse_items - The elements from this collection.

$sel:nextToken:ListConfigurationProfiles', listConfigurationProfilesResponse_nextToken - The token for the next set of items to return. Use this token to get the next set of results.

$sel:httpStatus:ListConfigurationProfilesResponse', listConfigurationProfilesResponse_httpStatus - The response's http status code.

Response Lenses

listConfigurationProfilesResponse_items :: Lens' ListConfigurationProfilesResponse (Maybe [ConfigurationProfileSummary])

The elements from this collection.

listConfigurationProfilesResponse_nextToken :: Lens' ListConfigurationProfilesResponse (Maybe Text)

The token for the next set of items to return. Use this token to get the next set of results.

listConfigurationProfilesResponse_httpStatus :: Lens' ListConfigurationProfilesResponse Int

The response's http status code.