Amazonka.AppConfig.StartDeployment

Description

Starts a deployment.

Creating a Request

data StartDeployment

See: newStartDeployment smart constructor.

Constructors

StartDeployment'

Fields

description :: Maybe Text

A description of the deployment.

tags :: Maybe (HashMap Text Text)

Metadata to assign to the deployment. Tags help organize and categorize your AppConfig resources. Each tag consists of a key and an optional value, both of which you define.

applicationId :: Text

The application ID.

environmentId :: Text

The environment ID.

deploymentStrategyId :: Text

The deployment strategy ID.

configurationProfileId :: Text

The configuration profile ID.

configurationVersion :: Text

The configuration version to deploy.

newStartDeployment

Arguments

:: Text
$sel:applicationId:StartDeployment'

-> Text
$sel:environmentId:StartDeployment'

-> Text
$sel:deploymentStrategyId:StartDeployment'

-> Text
$sel:configurationProfileId:StartDeployment'

-> Text
$sel:configurationVersion:StartDeployment'

-> StartDeployment

Create a value of StartDeployment with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:description:StartDeployment', startDeployment_description - A description of the deployment.

$sel:tags:StartDeployment', startDeployment_tags - Metadata to assign to the deployment. Tags help organize and categorize your AppConfig resources. Each tag consists of a key and an optional value, both of which you define.

$sel:applicationId:StartDeployment', startDeployment_applicationId - The application ID.

$sel:environmentId:StartDeployment', startDeployment_environmentId - The environment ID.

$sel:deploymentStrategyId:StartDeployment', startDeployment_deploymentStrategyId - The deployment strategy ID.

$sel:configurationProfileId:StartDeployment', startDeployment_configurationProfileId - The configuration profile ID.

$sel:configurationVersion:StartDeployment', startDeployment_configurationVersion - The configuration version to deploy.

Request Lenses

startDeployment_description :: Lens' StartDeployment (Maybe Text)

A description of the deployment.

startDeployment_tags :: Lens' StartDeployment (Maybe (HashMap Text Text))

Metadata to assign to the deployment. Tags help organize and categorize your AppConfig resources. Each tag consists of a key and an optional value, both of which you define.

startDeployment_applicationId :: Lens' StartDeployment Text

The application ID.

startDeployment_environmentId :: Lens' StartDeployment Text

The environment ID.

startDeployment_deploymentStrategyId :: Lens' StartDeployment Text

The deployment strategy ID.

startDeployment_configurationProfileId :: Lens' StartDeployment Text

The configuration profile ID.

startDeployment_configurationVersion :: Lens' StartDeployment Text

The configuration version to deploy.

Destructuring the Response

data Deployment

See: newDeployment smart constructor.

Constructors

Deployment'

Fields

growthFactor :: Maybe Double

The percentage of targets to receive a deployed configuration during each interval.

configurationName :: Maybe Text

The name of the configuration.

state :: Maybe DeploymentState

The state of the deployment.

deploymentStrategyId :: Maybe Text

The ID of the deployment strategy that was deployed.

deploymentNumber :: Maybe Int

The sequence number of the deployment.

configurationVersion :: Maybe Text

The configuration version that was deployed.

eventLog :: Maybe [DeploymentEvent]

A list containing all events related to a deployment. The most recent events are displayed first.

percentageComplete :: Maybe Double

The percentage of targets for which the deployment is available.

startedAt :: Maybe POSIX

The time the deployment started.

applicationId :: Maybe Text

The ID of the application that was deployed.

deploymentDurationInMinutes :: Maybe Natural

Total amount of time the deployment lasted.

environmentId :: Maybe Text

The ID of the environment that was deployed.

completedAt :: Maybe POSIX

The time the deployment completed.

configurationLocationUri :: Maybe Text

Information about the source location of the configuration.

finalBakeTimeInMinutes :: Maybe Natural

The amount of time AppConfig monitored for alarms before considering the deployment to be complete and no longer eligible for automatic roll back.

description :: Maybe Text

The description of the deployment.

configurationProfileId :: Maybe Text

The ID of the configuration profile that was deployed.

growthType :: Maybe GrowthType

The algorithm used to define how percentage grew over time.

newDeployment :: Deployment

Create a value of Deployment with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:growthFactor:Deployment', deployment_growthFactor - The percentage of targets to receive a deployed configuration during each interval.

$sel:configurationName:Deployment', deployment_configurationName - The name of the configuration.

$sel:state:Deployment', deployment_state - The state of the deployment.

$sel:deploymentStrategyId:Deployment', deployment_deploymentStrategyId - The ID of the deployment strategy that was deployed.

$sel:deploymentNumber:Deployment', deployment_deploymentNumber - The sequence number of the deployment.

$sel:configurationVersion:Deployment', deployment_configurationVersion - The configuration version that was deployed.

$sel:eventLog:Deployment', deployment_eventLog - A list containing all events related to a deployment. The most recent events are displayed first.

$sel:percentageComplete:Deployment', deployment_percentageComplete - The percentage of targets for which the deployment is available.

$sel:startedAt:Deployment', deployment_startedAt - The time the deployment started.

$sel:applicationId:Deployment', deployment_applicationId - The ID of the application that was deployed.

$sel:deploymentDurationInMinutes:Deployment', deployment_deploymentDurationInMinutes - Total amount of time the deployment lasted.

$sel:environmentId:Deployment', deployment_environmentId - The ID of the environment that was deployed.

$sel:completedAt:Deployment', deployment_completedAt - The time the deployment completed.

$sel:configurationLocationUri:Deployment', deployment_configurationLocationUri - Information about the source location of the configuration.

$sel:finalBakeTimeInMinutes:Deployment', deployment_finalBakeTimeInMinutes - The amount of time AppConfig monitored for alarms before considering the deployment to be complete and no longer eligible for automatic roll back.

$sel:description:Deployment', deployment_description - The description of the deployment.

$sel:configurationProfileId:Deployment', deployment_configurationProfileId - The ID of the configuration profile that was deployed.

$sel:growthType:Deployment', deployment_growthType - The algorithm used to define how percentage grew over time.

Response Lenses

deployment_growthFactor :: Lens' Deployment (Maybe Double)

The percentage of targets to receive a deployed configuration during each interval.

deployment_configurationName :: Lens' Deployment (Maybe Text)

The name of the configuration.

deployment_state :: Lens' Deployment (Maybe DeploymentState)

The state of the deployment.

deployment_deploymentStrategyId :: Lens' Deployment (Maybe Text)

The ID of the deployment strategy that was deployed.

deployment_deploymentNumber :: Lens' Deployment (Maybe Int)

The sequence number of the deployment.

deployment_configurationVersion :: Lens' Deployment (Maybe Text)

The configuration version that was deployed.

deployment_eventLog :: Lens' Deployment (Maybe [DeploymentEvent])

A list containing all events related to a deployment. The most recent events are displayed first.

deployment_percentageComplete :: Lens' Deployment (Maybe Double)

The percentage of targets for which the deployment is available.

deployment_startedAt :: Lens' Deployment (Maybe UTCTime)

The time the deployment started.

deployment_applicationId :: Lens' Deployment (Maybe Text)

The ID of the application that was deployed.

deployment_deploymentDurationInMinutes :: Lens' Deployment (Maybe Natural)

Total amount of time the deployment lasted.

deployment_environmentId :: Lens' Deployment (Maybe Text)

The ID of the environment that was deployed.

deployment_completedAt :: Lens' Deployment (Maybe UTCTime)

The time the deployment completed.

deployment_configurationLocationUri :: Lens' Deployment (Maybe Text)

Information about the source location of the configuration.

deployment_finalBakeTimeInMinutes :: Lens' Deployment (Maybe Natural)

The amount of time AppConfig monitored for alarms before considering the deployment to be complete and no longer eligible for automatic roll back.

deployment_description :: Lens' Deployment (Maybe Text)

The description of the deployment.

deployment_configurationProfileId :: Lens' Deployment (Maybe Text)

The ID of the configuration profile that was deployed.

deployment_growthType :: Lens' Deployment (Maybe GrowthType)

The algorithm used to define how percentage grew over time.