Amazonka.AppConfig.Types.ConfigurationProfile

Description

Documentation

data ConfigurationProfile

See: newConfigurationProfile smart constructor.

Constructors

ConfigurationProfile'

Fields

retrievalRoleArn :: Maybe Text

The ARN of an IAM role with permission to access the configuration at the specified LocationUri.

validators :: Maybe [Validator]

A list of methods for validating the configuration.

locationUri :: Maybe Text

The URI location of the configuration.

applicationId :: Maybe Text

The application ID.

name :: Maybe Text

The name of the configuration profile.

id :: Maybe Text

The configuration profile ID.

description :: Maybe Text

The configuration profile description.

newConfigurationProfile :: ConfigurationProfile

Create a value of ConfigurationProfile with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:retrievalRoleArn:ConfigurationProfile', configurationProfile_retrievalRoleArn - The ARN of an IAM role with permission to access the configuration at the specified LocationUri.

$sel:validators:ConfigurationProfile', configurationProfile_validators - A list of methods for validating the configuration.

$sel:locationUri:ConfigurationProfile', configurationProfile_locationUri - The URI location of the configuration.

$sel:applicationId:ConfigurationProfile', configurationProfile_applicationId - The application ID.

$sel:name:ConfigurationProfile', configurationProfile_name - The name of the configuration profile.

$sel:id:ConfigurationProfile', configurationProfile_id - The configuration profile ID.

$sel:description:ConfigurationProfile', configurationProfile_description - The configuration profile description.

configurationProfile_retrievalRoleArn :: Lens' ConfigurationProfile (Maybe Text)

The ARN of an IAM role with permission to access the configuration at the specified LocationUri.

configurationProfile_validators :: Lens' ConfigurationProfile (Maybe [Validator])

A list of methods for validating the configuration.

configurationProfile_locationUri :: Lens' ConfigurationProfile (Maybe Text)

The URI location of the configuration.

configurationProfile_applicationId :: Lens' ConfigurationProfile (Maybe Text)

The application ID.

configurationProfile_name :: Lens' ConfigurationProfile (Maybe Text)

The name of the configuration profile.

configurationProfile_id :: Lens' ConfigurationProfile (Maybe Text)

The configuration profile ID.

configurationProfile_description :: Lens' ConfigurationProfile (Maybe Text)

The configuration profile description.