Amazonka.AppConfig.Types.DeploymentSummary

Description

Documentation

data DeploymentSummary

Information about the deployment.

See: newDeploymentSummary smart constructor.

Constructors

DeploymentSummary'

Fields

growthFactor :: Maybe Double

The percentage of targets to receive a deployed configuration during each interval.

configurationName :: Maybe Text

The name of the configuration.

state :: Maybe DeploymentState

The state of the deployment.

deploymentNumber :: Maybe Int

The sequence number of the deployment.

configurationVersion :: Maybe Text

The version of the configuration.

percentageComplete :: Maybe Double

The percentage of targets for which the deployment is available.

startedAt :: Maybe POSIX

Time the deployment started.

deploymentDurationInMinutes :: Maybe Natural

Total amount of time the deployment lasted.

completedAt :: Maybe POSIX

Time the deployment completed.

finalBakeTimeInMinutes :: Maybe Natural

The amount of time AppConfig monitors for alarms before considering the deployment to be complete and no longer eligible for automatic roll back.

growthType :: Maybe GrowthType

The algorithm used to define how percentage grows over time.

newDeploymentSummary :: DeploymentSummary

Create a value of DeploymentSummary with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:growthFactor:DeploymentSummary', deploymentSummary_growthFactor - The percentage of targets to receive a deployed configuration during each interval.

$sel:configurationName:DeploymentSummary', deploymentSummary_configurationName - The name of the configuration.

$sel:state:DeploymentSummary', deploymentSummary_state - The state of the deployment.

$sel:deploymentNumber:DeploymentSummary', deploymentSummary_deploymentNumber - The sequence number of the deployment.

$sel:configurationVersion:DeploymentSummary', deploymentSummary_configurationVersion - The version of the configuration.

$sel:percentageComplete:DeploymentSummary', deploymentSummary_percentageComplete - The percentage of targets for which the deployment is available.

$sel:startedAt:DeploymentSummary', deploymentSummary_startedAt - Time the deployment started.

$sel:deploymentDurationInMinutes:DeploymentSummary', deploymentSummary_deploymentDurationInMinutes - Total amount of time the deployment lasted.

$sel:completedAt:DeploymentSummary', deploymentSummary_completedAt - Time the deployment completed.

$sel:finalBakeTimeInMinutes:DeploymentSummary', deploymentSummary_finalBakeTimeInMinutes - The amount of time AppConfig monitors for alarms before considering the deployment to be complete and no longer eligible for automatic roll back.

$sel:growthType:DeploymentSummary', deploymentSummary_growthType - The algorithm used to define how percentage grows over time.

deploymentSummary_growthFactor :: Lens' DeploymentSummary (Maybe Double)

The percentage of targets to receive a deployed configuration during each interval.

deploymentSummary_configurationName :: Lens' DeploymentSummary (Maybe Text)

The name of the configuration.

deploymentSummary_state :: Lens' DeploymentSummary (Maybe DeploymentState)

The state of the deployment.

deploymentSummary_deploymentNumber :: Lens' DeploymentSummary (Maybe Int)

The sequence number of the deployment.

deploymentSummary_configurationVersion :: Lens' DeploymentSummary (Maybe Text)

The version of the configuration.

deploymentSummary_percentageComplete :: Lens' DeploymentSummary (Maybe Double)

The percentage of targets for which the deployment is available.

deploymentSummary_startedAt :: Lens' DeploymentSummary (Maybe UTCTime)

Time the deployment started.

deploymentSummary_deploymentDurationInMinutes :: Lens' DeploymentSummary (Maybe Natural)

Total amount of time the deployment lasted.

deploymentSummary_completedAt :: Lens' DeploymentSummary (Maybe UTCTime)

Time the deployment completed.

deploymentSummary_finalBakeTimeInMinutes :: Lens' DeploymentSummary (Maybe Natural)

The amount of time AppConfig monitors for alarms before considering the deployment to be complete and no longer eligible for automatic roll back.

deploymentSummary_growthType :: Lens' DeploymentSummary (Maybe GrowthType)

The algorithm used to define how percentage grows over time.