Amazonka.AppConfig.Types.Environment

Description

Documentation

data Environment

See: newEnvironment smart constructor.

Constructors

Environment'

Fields

state :: Maybe EnvironmentState

The state of the environment. An environment can be in one of the following states: READY_FOR_DEPLOYMENT, DEPLOYING, ROLLING_BACK, or ROLLED_BACK

monitors :: Maybe [Monitor]

Amazon CloudWatch alarms monitored during the deployment.

applicationId :: Maybe Text

The application ID.

name :: Maybe Text

The name of the environment.

id :: Maybe Text

The environment ID.

description :: Maybe Text

The description of the environment.

newEnvironment :: Environment

Create a value of Environment with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:state:Environment', environment_state - The state of the environment. An environment can be in one of the following states: READY_FOR_DEPLOYMENT, DEPLOYING, ROLLING_BACK, or ROLLED_BACK

$sel:monitors:Environment', environment_monitors - Amazon CloudWatch alarms monitored during the deployment.

$sel:applicationId:Environment', environment_applicationId - The application ID.

$sel:name:Environment', environment_name - The name of the environment.

$sel:id:Environment', environment_id - The environment ID.

$sel:description:Environment', environment_description - The description of the environment.

environment_state :: Lens' Environment (Maybe EnvironmentState)

The state of the environment. An environment can be in one of the following states: READY_FOR_DEPLOYMENT, DEPLOYING, ROLLING_BACK, or ROLLED_BACK

environment_monitors :: Lens' Environment (Maybe [Monitor])

Amazon CloudWatch alarms monitored during the deployment.

environment_applicationId :: Lens' Environment (Maybe Text)

The application ID.

environment_name :: Lens' Environment (Maybe Text)

The name of the environment.

environment_id :: Lens' Environment (Maybe Text)

The environment ID.

environment_description :: Lens' Environment (Maybe Text)

The description of the environment.