Amazonka.AppConfig.Types.HostedConfigurationVersion

Description

Documentation

data HostedConfigurationVersion

See: newHostedConfigurationVersion smart constructor.

Constructors

HostedConfigurationVersion'

Fields

content :: Maybe (Sensitive ByteString)

The content of the configuration or the configuration data.

versionNumber :: Maybe Int

The configuration version.

applicationId :: Maybe Text

The application ID.

description :: Maybe Text

A description of the configuration.

configurationProfileId :: Maybe Text

The configuration profile ID.

contentType :: Maybe Text

A standard MIME type describing the format of the configuration content. For more information, see Content-Type.

newHostedConfigurationVersion :: HostedConfigurationVersion

Create a value of HostedConfigurationVersion with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:content:HostedConfigurationVersion', hostedConfigurationVersion_content - The content of the configuration or the configuration data.

$sel:versionNumber:HostedConfigurationVersion', hostedConfigurationVersion_versionNumber - The configuration version.

$sel:applicationId:HostedConfigurationVersion', hostedConfigurationVersion_applicationId - The application ID.

$sel:description:HostedConfigurationVersion', hostedConfigurationVersion_description - A description of the configuration.

$sel:configurationProfileId:HostedConfigurationVersion', hostedConfigurationVersion_configurationProfileId - The configuration profile ID.

$sel:contentType:HostedConfigurationVersion', hostedConfigurationVersion_contentType - A standard MIME type describing the format of the configuration content. For more information, see Content-Type.

hostedConfigurationVersion_content :: Lens' HostedConfigurationVersion (Maybe ByteString)

The content of the configuration or the configuration data.

hostedConfigurationVersion_versionNumber :: Lens' HostedConfigurationVersion (Maybe Int)

The configuration version.

hostedConfigurationVersion_applicationId :: Lens' HostedConfigurationVersion (Maybe Text)

The application ID.

hostedConfigurationVersion_description :: Lens' HostedConfigurationVersion (Maybe Text)

A description of the configuration.

hostedConfigurationVersion_configurationProfileId :: Lens' HostedConfigurationVersion (Maybe Text)

The configuration profile ID.

hostedConfigurationVersion_contentType :: Lens' HostedConfigurationVersion (Maybe Text)

A standard MIME type describing the format of the configuration content. For more information, see Content-Type.