Amazonka.AppConfig.UpdateApplication

Description

Updates an application.

Creating a Request

data UpdateApplication

See: newUpdateApplication smart constructor.

Constructors

UpdateApplication'

Fields

name :: Maybe Text

The name of the application.

description :: Maybe Text

A description of the application.

applicationId :: Text

The application ID.

newUpdateApplication

Arguments

:: Text
$sel:applicationId:UpdateApplication'

-> UpdateApplication

Create a value of UpdateApplication with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:name:UpdateApplication', updateApplication_name - The name of the application.

$sel:description:UpdateApplication', updateApplication_description - A description of the application.

$sel:applicationId:UpdateApplication', updateApplication_applicationId - The application ID.

Request Lenses

updateApplication_name :: Lens' UpdateApplication (Maybe Text)

The name of the application.

updateApplication_description :: Lens' UpdateApplication (Maybe Text)

A description of the application.

updateApplication_applicationId :: Lens' UpdateApplication Text

The application ID.

Destructuring the Response

data Application

See: newApplication smart constructor.

Constructors

Application'

Fields

name :: Maybe Text

The application name.

id :: Maybe Text

The application ID.

description :: Maybe Text

The description of the application.

newApplication :: Application

Create a value of Application with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:name:Application', application_name - The application name.

$sel:id:Application', application_id - The application ID.

$sel:description:Application', application_description - The description of the application.

Response Lenses

application_name :: Lens' Application (Maybe Text)

The application name.

application_id :: Lens' Application (Maybe Text)

The application ID.

application_description :: Lens' Application (Maybe Text)

The description of the application.