Amazonka.AppFlow.CreateConnectorProfile

Description

Creates a new connector profile associated with your Amazon Web Services account. There is a soft quota of 100 connector profiles per Amazon Web Services account. If you need more connector profiles than this quota allows, you can submit a request to the Amazon AppFlow team through the Amazon AppFlow support channel.

Creating a Request

data CreateConnectorProfile

See: newCreateConnectorProfile smart constructor.

Constructors

CreateConnectorProfile'

Fields

kmsArn :: Maybe Text

The ARN (Amazon Resource Name) of the Key Management Service (KMS) key you provide for encryption. This is required if you do not want to use the Amazon AppFlow-managed KMS key. If you don't provide anything here, Amazon AppFlow uses the Amazon AppFlow-managed KMS key.

connectorProfileName :: Text

The name of the connector profile. The name is unique for each ConnectorProfile in your Amazon Web Services account.

connectorType :: ConnectorType

The type of connector, such as Salesforce, Amplitude, and so on.

connectionMode :: ConnectionMode

Indicates the connection mode and specifies whether it is public or private. Private flows use Amazon Web Services PrivateLink to route data over Amazon Web Services infrastructure without exposing it to the public internet.

connectorProfileConfig :: ConnectorProfileConfig

Defines the connector-specific configuration and credentials.

newCreateConnectorProfile

Arguments

:: Text
$sel:connectorProfileName:CreateConnectorProfile'

-> ConnectorType
$sel:connectorType:CreateConnectorProfile'

-> ConnectionMode
$sel:connectionMode:CreateConnectorProfile'

-> ConnectorProfileConfig
$sel:connectorProfileConfig:CreateConnectorProfile'

-> CreateConnectorProfile

Create a value of CreateConnectorProfile with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:kmsArn:CreateConnectorProfile', createConnectorProfile_kmsArn - The ARN (Amazon Resource Name) of the Key Management Service (KMS) key you provide for encryption. This is required if you do not want to use the Amazon AppFlow-managed KMS key. If you don't provide anything here, Amazon AppFlow uses the Amazon AppFlow-managed KMS key.

$sel:connectorProfileName:CreateConnectorProfile', createConnectorProfile_connectorProfileName - The name of the connector profile. The name is unique for each ConnectorProfile in your Amazon Web Services account.

$sel:connectorType:CreateConnectorProfile', createConnectorProfile_connectorType - The type of connector, such as Salesforce, Amplitude, and so on.

$sel:connectionMode:CreateConnectorProfile', createConnectorProfile_connectionMode - Indicates the connection mode and specifies whether it is public or private. Private flows use Amazon Web Services PrivateLink to route data over Amazon Web Services infrastructure without exposing it to the public internet.

$sel:connectorProfileConfig:CreateConnectorProfile', createConnectorProfile_connectorProfileConfig - Defines the connector-specific configuration and credentials.

Request Lenses

createConnectorProfile_kmsArn :: Lens' CreateConnectorProfile (Maybe Text)

The ARN (Amazon Resource Name) of the Key Management Service (KMS) key you provide for encryption. This is required if you do not want to use the Amazon AppFlow-managed KMS key. If you don't provide anything here, Amazon AppFlow uses the Amazon AppFlow-managed KMS key.

createConnectorProfile_connectorProfileName :: Lens' CreateConnectorProfile Text

The name of the connector profile. The name is unique for each ConnectorProfile in your Amazon Web Services account.

createConnectorProfile_connectorType :: Lens' CreateConnectorProfile ConnectorType

The type of connector, such as Salesforce, Amplitude, and so on.

createConnectorProfile_connectionMode :: Lens' CreateConnectorProfile ConnectionMode

Indicates the connection mode and specifies whether it is public or private. Private flows use Amazon Web Services PrivateLink to route data over Amazon Web Services infrastructure without exposing it to the public internet.

createConnectorProfile_connectorProfileConfig :: Lens' CreateConnectorProfile ConnectorProfileConfig

Defines the connector-specific configuration and credentials.

Destructuring the Response

data CreateConnectorProfileResponse

See: newCreateConnectorProfileResponse smart constructor.

Constructors

CreateConnectorProfileResponse'

Fields

connectorProfileArn :: Maybe Text

The Amazon Resource Name (ARN) of the connector profile.

httpStatus :: Int

The response's http status code.

newCreateConnectorProfileResponse

Arguments

:: Int
$sel:httpStatus:CreateConnectorProfileResponse'

-> CreateConnectorProfileResponse

Create a value of CreateConnectorProfileResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:connectorProfileArn:CreateConnectorProfileResponse', createConnectorProfileResponse_connectorProfileArn - The Amazon Resource Name (ARN) of the connector profile.

$sel:httpStatus:CreateConnectorProfileResponse', createConnectorProfileResponse_httpStatus - The response's http status code.

Response Lenses

createConnectorProfileResponse_connectorProfileArn :: Lens' CreateConnectorProfileResponse (Maybe Text)

The Amazon Resource Name (ARN) of the connector profile.

createConnectorProfileResponse_httpStatus :: Lens' CreateConnectorProfileResponse Int

The response's http status code.