Amazonka.AppFlow.CreateFlow

Description

Enables your application to create a new flow using Amazon AppFlow. You must create a connector profile before calling this API. Please note that the Request Syntax below shows syntax for multiple destinations, however, you can only transfer data to one item in this list at a time. Amazon AppFlow does not currently support flows to multiple destinations at once.

Creating a Request

data CreateFlow

See: newCreateFlow smart constructor.

Constructors

CreateFlow'

Fields

kmsArn :: Maybe Text

The ARN (Amazon Resource Name) of the Key Management Service (KMS) key you provide for encryption. This is required if you do not want to use the Amazon AppFlow-managed KMS key. If you don't provide anything here, Amazon AppFlow uses the Amazon AppFlow-managed KMS key.

description :: Maybe Text

A description of the flow you want to create.

tags :: Maybe (HashMap Text Text)

The tags used to organize, track, or control access for your flow.

flowName :: Text

The specified name of the flow. Spaces are not allowed. Use underscores (_) or hyphens (-) only.

triggerConfig :: TriggerConfig

The trigger settings that determine how and when the flow runs.

sourceFlowConfig :: SourceFlowConfig

The configuration that controls how Amazon AppFlow retrieves data from the source connector.

destinationFlowConfigList :: [DestinationFlowConfig]

The configuration that controls how Amazon AppFlow places data in the destination connector.

tasks :: [Task]

A list of tasks that Amazon AppFlow performs while transferring the data in the flow run.

newCreateFlow

Arguments

:: Text
$sel:flowName:CreateFlow'

-> TriggerConfig
$sel:triggerConfig:CreateFlow'

-> SourceFlowConfig
$sel:sourceFlowConfig:CreateFlow'

-> CreateFlow

Create a value of CreateFlow with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:kmsArn:CreateFlow', createFlow_kmsArn - The ARN (Amazon Resource Name) of the Key Management Service (KMS) key you provide for encryption. This is required if you do not want to use the Amazon AppFlow-managed KMS key. If you don't provide anything here, Amazon AppFlow uses the Amazon AppFlow-managed KMS key.

$sel:description:CreateFlow', createFlow_description - A description of the flow you want to create.

$sel:tags:CreateFlow', createFlow_tags - The tags used to organize, track, or control access for your flow.

$sel:flowName:CreateFlow', createFlow_flowName - The specified name of the flow. Spaces are not allowed. Use underscores (_) or hyphens (-) only.

$sel:triggerConfig:CreateFlow', createFlow_triggerConfig - The trigger settings that determine how and when the flow runs.

$sel:sourceFlowConfig:CreateFlow', createFlow_sourceFlowConfig - The configuration that controls how Amazon AppFlow retrieves data from the source connector.

$sel:destinationFlowConfigList:CreateFlow', createFlow_destinationFlowConfigList - The configuration that controls how Amazon AppFlow places data in the destination connector.

$sel:tasks:CreateFlow', createFlow_tasks - A list of tasks that Amazon AppFlow performs while transferring the data in the flow run.

Request Lenses

createFlow_kmsArn :: Lens' CreateFlow (Maybe Text)

The ARN (Amazon Resource Name) of the Key Management Service (KMS) key you provide for encryption. This is required if you do not want to use the Amazon AppFlow-managed KMS key. If you don't provide anything here, Amazon AppFlow uses the Amazon AppFlow-managed KMS key.

createFlow_description :: Lens' CreateFlow (Maybe Text)

A description of the flow you want to create.

createFlow_tags :: Lens' CreateFlow (Maybe (HashMap Text Text))

The tags used to organize, track, or control access for your flow.

createFlow_flowName :: Lens' CreateFlow Text

The specified name of the flow. Spaces are not allowed. Use underscores (_) or hyphens (-) only.

createFlow_triggerConfig :: Lens' CreateFlow TriggerConfig

The trigger settings that determine how and when the flow runs.

createFlow_sourceFlowConfig :: Lens' CreateFlow SourceFlowConfig

The configuration that controls how Amazon AppFlow retrieves data from the source connector.

createFlow_destinationFlowConfigList :: Lens' CreateFlow [DestinationFlowConfig]

The configuration that controls how Amazon AppFlow places data in the destination connector.

createFlow_tasks :: Lens' CreateFlow [Task]

A list of tasks that Amazon AppFlow performs while transferring the data in the flow run.

Destructuring the Response

data CreateFlowResponse

See: newCreateFlowResponse smart constructor.

Constructors

CreateFlowResponse'

Fields

flowArn :: Maybe Text

The flow's Amazon Resource Name (ARN).

flowStatus :: Maybe FlowStatus

Indicates the current status of the flow.

httpStatus :: Int

The response's http status code.

newCreateFlowResponse

Arguments

:: Int
$sel:httpStatus:CreateFlowResponse'

-> CreateFlowResponse

Create a value of CreateFlowResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:flowArn:CreateFlowResponse', createFlowResponse_flowArn - The flow's Amazon Resource Name (ARN).

$sel:flowStatus:CreateFlowResponse', createFlowResponse_flowStatus - Indicates the current status of the flow.

$sel:httpStatus:CreateFlowResponse', createFlowResponse_httpStatus - The response's http status code.

Response Lenses

createFlowResponse_flowArn :: Lens' CreateFlowResponse (Maybe Text)

The flow's Amazon Resource Name (ARN).

createFlowResponse_flowStatus :: Lens' CreateFlowResponse (Maybe FlowStatus)

Indicates the current status of the flow.

createFlowResponse_httpStatus :: Lens' CreateFlowResponse Int

The response's http status code.