Amazonka.AppFlow.ListConnectorEntities

Description

Returns the list of available connector entities supported by Amazon AppFlow. For example, you can query Salesforce for Account and Opportunity entities, or query ServiceNow for the Incident entity.

Creating a Request

data ListConnectorEntities

See: newListConnectorEntities smart constructor.

Constructors

ListConnectorEntities'

Fields

connectorProfileName :: Maybe Text

The name of the connector profile. The name is unique for each ConnectorProfile in the Amazon Web Services account, and is used to query the downstream connector.

entitiesPath :: Maybe Text

This optional parameter is specific to connector implementation. Some connectors support multiple levels or categories of entities. You can find out the list of roots for such providers by sending a request without the entitiesPath parameter. If the connector supports entities at different roots, this initial request returns the list of roots. Otherwise, this request returns all entities supported by the provider.

connectorType :: Maybe ConnectorType

The type of connector, such as Salesforce, Amplitude, and so on.

newListConnectorEntities :: ListConnectorEntities

Create a value of ListConnectorEntities with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:connectorProfileName:ListConnectorEntities', listConnectorEntities_connectorProfileName - The name of the connector profile. The name is unique for each ConnectorProfile in the Amazon Web Services account, and is used to query the downstream connector.

$sel:entitiesPath:ListConnectorEntities', listConnectorEntities_entitiesPath - This optional parameter is specific to connector implementation. Some connectors support multiple levels or categories of entities. You can find out the list of roots for such providers by sending a request without the entitiesPath parameter. If the connector supports entities at different roots, this initial request returns the list of roots. Otherwise, this request returns all entities supported by the provider.

$sel:connectorType:ListConnectorEntities', listConnectorEntities_connectorType - The type of connector, such as Salesforce, Amplitude, and so on.

Request Lenses

listConnectorEntities_connectorProfileName :: Lens' ListConnectorEntities (Maybe Text)

The name of the connector profile. The name is unique for each ConnectorProfile in the Amazon Web Services account, and is used to query the downstream connector.

listConnectorEntities_entitiesPath :: Lens' ListConnectorEntities (Maybe Text)

This optional parameter is specific to connector implementation. Some connectors support multiple levels or categories of entities. You can find out the list of roots for such providers by sending a request without the entitiesPath parameter. If the connector supports entities at different roots, this initial request returns the list of roots. Otherwise, this request returns all entities supported by the provider.

listConnectorEntities_connectorType :: Lens' ListConnectorEntities (Maybe ConnectorType)

The type of connector, such as Salesforce, Amplitude, and so on.

Destructuring the Response

data ListConnectorEntitiesResponse

See: newListConnectorEntitiesResponse smart constructor.

Constructors

ListConnectorEntitiesResponse'

Fields

httpStatus :: Int

The response's http status code.

connectorEntityMap :: HashMap Text [ConnectorEntity]

The response of ListConnectorEntities lists entities grouped by category. This map's key represents the group name, and its value contains the list of entities belonging to that group.

newListConnectorEntitiesResponse

Arguments

:: Int
$sel:httpStatus:ListConnectorEntitiesResponse'

-> ListConnectorEntitiesResponse

Create a value of ListConnectorEntitiesResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:httpStatus:ListConnectorEntitiesResponse', listConnectorEntitiesResponse_httpStatus - The response's http status code.

$sel:connectorEntityMap:ListConnectorEntitiesResponse', listConnectorEntitiesResponse_connectorEntityMap - The response of ListConnectorEntities lists entities grouped by category. This map's key represents the group name, and its value contains the list of entities belonging to that group.

Response Lenses

listConnectorEntitiesResponse_httpStatus :: Lens' ListConnectorEntitiesResponse Int

The response's http status code.

listConnectorEntitiesResponse_connectorEntityMap :: Lens' ListConnectorEntitiesResponse (HashMap Text [ConnectorEntity])

The response of ListConnectorEntities lists entities grouped by category. This map's key represents the group name, and its value contains the list of entities belonging to that group.