Amazonka.AppFlow.Types.ConnectorEntityField

Description

Documentation

data ConnectorEntityField

Describes the data model of a connector field. For example, for an account entity, the fields would be account name, account ID, and so on.

See: newConnectorEntityField smart constructor.

Constructors

ConnectorEntityField'

Fields

sourceProperties :: Maybe SourceFieldProperties

The properties that can be applied to a field when the connector is being used as a source.

supportedFieldTypeDetails :: Maybe SupportedFieldTypeDetails

Contains details regarding the supported FieldType, including the corresponding filterOperators and supportedValues.

destinationProperties :: Maybe DestinationFieldProperties

The properties applied to a field when the connector is being used as a destination.

description :: Maybe Text

A description of the connector entity field.

label :: Maybe Text

The label applied to a connector entity field.

identifier :: Text

The unique identifier of the connector field.

newConnectorEntityField

Arguments

:: Text
$sel:identifier:ConnectorEntityField'

-> ConnectorEntityField

Create a value of ConnectorEntityField with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:sourceProperties:ConnectorEntityField', connectorEntityField_sourceProperties - The properties that can be applied to a field when the connector is being used as a source.

$sel:supportedFieldTypeDetails:ConnectorEntityField', connectorEntityField_supportedFieldTypeDetails - Contains details regarding the supported FieldType, including the corresponding filterOperators and supportedValues.

$sel:destinationProperties:ConnectorEntityField', connectorEntityField_destinationProperties - The properties applied to a field when the connector is being used as a destination.

$sel:description:ConnectorEntityField', connectorEntityField_description - A description of the connector entity field.

$sel:label:ConnectorEntityField', connectorEntityField_label - The label applied to a connector entity field.

$sel:identifier:ConnectorEntityField', connectorEntityField_identifier - The unique identifier of the connector field.

connectorEntityField_sourceProperties :: Lens' ConnectorEntityField (Maybe SourceFieldProperties)

The properties that can be applied to a field when the connector is being used as a source.

connectorEntityField_supportedFieldTypeDetails :: Lens' ConnectorEntityField (Maybe SupportedFieldTypeDetails)

Contains details regarding the supported FieldType, including the corresponding filterOperators and supportedValues.

connectorEntityField_destinationProperties :: Lens' ConnectorEntityField (Maybe DestinationFieldProperties)

The properties applied to a field when the connector is being used as a destination.

connectorEntityField_description :: Lens' ConnectorEntityField (Maybe Text)

A description of the connector entity field.

connectorEntityField_label :: Lens' ConnectorEntityField (Maybe Text)

The label applied to a connector entity field.

connectorEntityField_identifier :: Lens' ConnectorEntityField Text

The unique identifier of the connector field.