Amazonka.AppFlow.Types.ConnectorProfile

Description

Documentation

data ConnectorProfile

Describes an instance of a connector. This includes the provided name, credentials ARN, connection-mode, and so on. To keep the API intuitive and extensible, the fields that are common to all types of connector profiles are explicitly specified at the top level. The rest of the connector-specific properties are available via the connectorProfileProperties field.

See: newConnectorProfile smart constructor.

Constructors

ConnectorProfile'

Fields

connectorProfileName :: Maybe Text

The name of the connector profile. The name is unique for each ConnectorProfile in the Amazon Web Services account.

lastUpdatedAt :: Maybe POSIX

Specifies when the connector profile was last updated.

createdAt :: Maybe POSIX

Specifies when the connector profile was created.

credentialsArn :: Maybe Text

The Amazon Resource Name (ARN) of the connector profile credentials.

connectorProfileProperties :: Maybe ConnectorProfileProperties

The connector-specific properties of the profile configuration.

connectionMode :: Maybe ConnectionMode

Indicates the connection mode and if it is public or private.

connectorProfileArn :: Maybe Text

The Amazon Resource Name (ARN) of the connector profile.

privateConnectionProvisioningState :: Maybe PrivateConnectionProvisioningState

Specifies the private connection provisioning state.

connectorType :: Maybe ConnectorType

The type of connector, such as Salesforce, Amplitude, and so on.

newConnectorProfile :: ConnectorProfile

Create a value of ConnectorProfile with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:connectorProfileName:ConnectorProfile', connectorProfile_connectorProfileName - The name of the connector profile. The name is unique for each ConnectorProfile in the Amazon Web Services account.

$sel:lastUpdatedAt:ConnectorProfile', connectorProfile_lastUpdatedAt - Specifies when the connector profile was last updated.

$sel:createdAt:ConnectorProfile', connectorProfile_createdAt - Specifies when the connector profile was created.

$sel:credentialsArn:ConnectorProfile', connectorProfile_credentialsArn - The Amazon Resource Name (ARN) of the connector profile credentials.

$sel:connectorProfileProperties:ConnectorProfile', connectorProfile_connectorProfileProperties - The connector-specific properties of the profile configuration.

$sel:connectionMode:ConnectorProfile', connectorProfile_connectionMode - Indicates the connection mode and if it is public or private.

$sel:connectorProfileArn:ConnectorProfile', connectorProfile_connectorProfileArn - The Amazon Resource Name (ARN) of the connector profile.

$sel:privateConnectionProvisioningState:ConnectorProfile', connectorProfile_privateConnectionProvisioningState - Specifies the private connection provisioning state.

$sel:connectorType:ConnectorProfile', connectorProfile_connectorType - The type of connector, such as Salesforce, Amplitude, and so on.

connectorProfile_connectorProfileName :: Lens' ConnectorProfile (Maybe Text)

The name of the connector profile. The name is unique for each ConnectorProfile in the Amazon Web Services account.

connectorProfile_lastUpdatedAt :: Lens' ConnectorProfile (Maybe UTCTime)

Specifies when the connector profile was last updated.

connectorProfile_createdAt :: Lens' ConnectorProfile (Maybe UTCTime)

Specifies when the connector profile was created.

connectorProfile_credentialsArn :: Lens' ConnectorProfile (Maybe Text)

The Amazon Resource Name (ARN) of the connector profile credentials.

connectorProfile_connectorProfileProperties :: Lens' ConnectorProfile (Maybe ConnectorProfileProperties)

The connector-specific properties of the profile configuration.

connectorProfile_connectionMode :: Lens' ConnectorProfile (Maybe ConnectionMode)

Indicates the connection mode and if it is public or private.

connectorProfile_connectorProfileArn :: Lens' ConnectorProfile (Maybe Text)

The Amazon Resource Name (ARN) of the connector profile.

connectorProfile_privateConnectionProvisioningState :: Lens' ConnectorProfile (Maybe PrivateConnectionProvisioningState)

Specifies the private connection provisioning state.

connectorProfile_connectorType :: Lens' ConnectorProfile (Maybe ConnectorType)

The type of connector, such as Salesforce, Amplitude, and so on.