Amazonka.AppFlow.Types.DestinationConnectorProperties

Description

Documentation

data DestinationConnectorProperties

This stores the information that is required to query a particular connector.

See: newDestinationConnectorProperties smart constructor.

Constructors

DestinationConnectorProperties'

Fields

upsolver :: Maybe UpsolverDestinationProperties

The properties required to query Upsolver.

snowflake :: Maybe SnowflakeDestinationProperties

The properties required to query Snowflake.

honeycode :: Maybe HoneycodeDestinationProperties

The properties required to query Amazon Honeycode.

lookoutMetrics :: Maybe LookoutMetricsDestinationProperties

The properties required to query Amazon Lookout for Metrics.

customerProfiles :: Maybe CustomerProfilesDestinationProperties

The properties required to query Amazon Connect Customer Profiles.

salesforce :: Maybe SalesforceDestinationProperties

The properties required to query Salesforce.

zendesk :: Maybe ZendeskDestinationProperties

The properties required to query Zendesk.

s3 :: Maybe S3DestinationProperties

The properties required to query Amazon S3.

eventBridge :: Maybe EventBridgeDestinationProperties

The properties required to query Amazon EventBridge.

redshift :: Maybe RedshiftDestinationProperties

The properties required to query Amazon Redshift.

newDestinationConnectorProperties :: DestinationConnectorProperties

Create a value of DestinationConnectorProperties with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:upsolver:DestinationConnectorProperties', destinationConnectorProperties_upsolver - The properties required to query Upsolver.

$sel:snowflake:DestinationConnectorProperties', destinationConnectorProperties_snowflake - The properties required to query Snowflake.

$sel:honeycode:DestinationConnectorProperties', destinationConnectorProperties_honeycode - The properties required to query Amazon Honeycode.

$sel:lookoutMetrics:DestinationConnectorProperties', destinationConnectorProperties_lookoutMetrics - The properties required to query Amazon Lookout for Metrics.

$sel:customerProfiles:DestinationConnectorProperties', destinationConnectorProperties_customerProfiles - The properties required to query Amazon Connect Customer Profiles.

$sel:salesforce:DestinationConnectorProperties', destinationConnectorProperties_salesforce - The properties required to query Salesforce.

$sel:zendesk:DestinationConnectorProperties', destinationConnectorProperties_zendesk - The properties required to query Zendesk.

$sel:s3:DestinationConnectorProperties', destinationConnectorProperties_s3 - The properties required to query Amazon S3.

$sel:eventBridge:DestinationConnectorProperties', destinationConnectorProperties_eventBridge - The properties required to query Amazon EventBridge.

$sel:redshift:DestinationConnectorProperties', destinationConnectorProperties_redshift - The properties required to query Amazon Redshift.

destinationConnectorProperties_upsolver :: Lens' DestinationConnectorProperties (Maybe UpsolverDestinationProperties)

The properties required to query Upsolver.

destinationConnectorProperties_snowflake :: Lens' DestinationConnectorProperties (Maybe SnowflakeDestinationProperties)

The properties required to query Snowflake.

destinationConnectorProperties_honeycode :: Lens' DestinationConnectorProperties (Maybe HoneycodeDestinationProperties)

The properties required to query Amazon Honeycode.

destinationConnectorProperties_lookoutMetrics :: Lens' DestinationConnectorProperties (Maybe LookoutMetricsDestinationProperties)

The properties required to query Amazon Lookout for Metrics.

destinationConnectorProperties_customerProfiles :: Lens' DestinationConnectorProperties (Maybe CustomerProfilesDestinationProperties)

The properties required to query Amazon Connect Customer Profiles.

destinationConnectorProperties_salesforce :: Lens' DestinationConnectorProperties (Maybe SalesforceDestinationProperties)

The properties required to query Salesforce.

destinationConnectorProperties_zendesk :: Lens' DestinationConnectorProperties (Maybe ZendeskDestinationProperties)

The properties required to query Zendesk.

destinationConnectorProperties_s3 :: Lens' DestinationConnectorProperties (Maybe S3DestinationProperties)

The properties required to query Amazon S3.

destinationConnectorProperties_eventBridge :: Lens' DestinationConnectorProperties (Maybe EventBridgeDestinationProperties)

The properties required to query Amazon EventBridge.

destinationConnectorProperties_redshift :: Lens' DestinationConnectorProperties (Maybe RedshiftDestinationProperties)

The properties required to query Amazon Redshift.