Amazonka.AppFlow.Types.DestinationFieldProperties

Description

Documentation

data DestinationFieldProperties

The properties that can be applied to a field when connector is being used as a destination.

See: newDestinationFieldProperties smart constructor.

Constructors

DestinationFieldProperties'

Fields

isUpdatable :: Maybe Bool

Specifies whether the field can be updated during an UPDATE or UPSERT write operation.

isNullable :: Maybe Bool

Specifies if the destination field can have a null value.

supportedWriteOperations :: Maybe [WriteOperationType]

A list of supported write operations. For each write operation listed, this field can be used in idFieldNames when that write operation is present as a destination option.

isUpsertable :: Maybe Bool

Specifies if the flow run can either insert new rows in the destination field if they do not already exist, or update them if they do.

isCreatable :: Maybe Bool

Specifies if the destination field can be created by the current user.

newDestinationFieldProperties :: DestinationFieldProperties

Create a value of DestinationFieldProperties with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:isUpdatable:DestinationFieldProperties', destinationFieldProperties_isUpdatable - Specifies whether the field can be updated during an UPDATE or UPSERT write operation.

$sel:isNullable:DestinationFieldProperties', destinationFieldProperties_isNullable - Specifies if the destination field can have a null value.

$sel:supportedWriteOperations:DestinationFieldProperties', destinationFieldProperties_supportedWriteOperations - A list of supported write operations. For each write operation listed, this field can be used in idFieldNames when that write operation is present as a destination option.

$sel:isUpsertable:DestinationFieldProperties', destinationFieldProperties_isUpsertable - Specifies if the flow run can either insert new rows in the destination field if they do not already exist, or update them if they do.

$sel:isCreatable:DestinationFieldProperties', destinationFieldProperties_isCreatable - Specifies if the destination field can be created by the current user.

destinationFieldProperties_isUpdatable :: Lens' DestinationFieldProperties (Maybe Bool)

Specifies whether the field can be updated during an UPDATE or UPSERT write operation.

destinationFieldProperties_isNullable :: Lens' DestinationFieldProperties (Maybe Bool)

Specifies if the destination field can have a null value.

destinationFieldProperties_supportedWriteOperations :: Lens' DestinationFieldProperties (Maybe [WriteOperationType])

A list of supported write operations. For each write operation listed, this field can be used in idFieldNames when that write operation is present as a destination option.

destinationFieldProperties_isUpsertable :: Lens' DestinationFieldProperties (Maybe Bool)

Specifies if the flow run can either insert new rows in the destination field if they do not already exist, or update them if they do.

destinationFieldProperties_isCreatable :: Lens' DestinationFieldProperties (Maybe Bool)

Specifies if the destination field can be created by the current user.