Amazonka.AppFlow.Types.ExecutionResult

Description

Documentation

data ExecutionResult

Specifies the end result of the flow run.

See: newExecutionResult smart constructor.

Constructors

ExecutionResult'

Fields

recordsProcessed :: Maybe Integer

The number of records processed in the flow run.

bytesWritten :: Maybe Integer

The total number of bytes written as a result of the flow run.

bytesProcessed :: Maybe Integer

The total number of bytes processed by the flow run.

errorInfo :: Maybe ErrorInfo

Provides any error message information related to the flow run.

newExecutionResult :: ExecutionResult

Create a value of ExecutionResult with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:recordsProcessed:ExecutionResult', executionResult_recordsProcessed - The number of records processed in the flow run.

$sel:bytesWritten:ExecutionResult', executionResult_bytesWritten - The total number of bytes written as a result of the flow run.

$sel:bytesProcessed:ExecutionResult', executionResult_bytesProcessed - The total number of bytes processed by the flow run.

$sel:errorInfo:ExecutionResult', executionResult_errorInfo - Provides any error message information related to the flow run.

executionResult_recordsProcessed :: Lens' ExecutionResult (Maybe Integer)

The number of records processed in the flow run.

executionResult_bytesWritten :: Lens' ExecutionResult (Maybe Integer)

The total number of bytes written as a result of the flow run.

executionResult_bytesProcessed :: Lens' ExecutionResult (Maybe Integer)

The total number of bytes processed by the flow run.

executionResult_errorInfo :: Lens' ExecutionResult (Maybe ErrorInfo)

Provides any error message information related to the flow run.