Amazonka.AppFlow.Types.GoogleAnalyticsConnectorProfileCredentials

Description

Documentation

data GoogleAnalyticsConnectorProfileCredentials

The connector-specific profile credentials required by Google Analytics.

See: newGoogleAnalyticsConnectorProfileCredentials smart constructor.

Constructors

GoogleAnalyticsConnectorProfileCredentials'

Fields

accessToken :: Maybe (Sensitive Text)

The credentials used to access protected Google Analytics resources.

refreshToken :: Maybe Text

The credentials used to acquire new access tokens. This is required only for OAuth2 access tokens, and is not required for OAuth1 access tokens.

oAuthRequest :: Maybe ConnectorOAuthRequest

The OAuth requirement needed to request security tokens from the connector endpoint.

clientId :: Text

The identifier for the desired client.

clientSecret :: Sensitive Text

The client secret used by the OAuth client to authenticate to the authorization server.

newGoogleAnalyticsConnectorProfileCredentials

Arguments

:: Text
$sel:clientId:GoogleAnalyticsConnectorProfileCredentials'

-> Text
$sel:clientSecret:GoogleAnalyticsConnectorProfileCredentials'

-> GoogleAnalyticsConnectorProfileCredentials

Create a value of GoogleAnalyticsConnectorProfileCredentials with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:accessToken:GoogleAnalyticsConnectorProfileCredentials', googleAnalyticsConnectorProfileCredentials_accessToken - The credentials used to access protected Google Analytics resources.

$sel:refreshToken:GoogleAnalyticsConnectorProfileCredentials', googleAnalyticsConnectorProfileCredentials_refreshToken - The credentials used to acquire new access tokens. This is required only for OAuth2 access tokens, and is not required for OAuth1 access tokens.

$sel:oAuthRequest:GoogleAnalyticsConnectorProfileCredentials', googleAnalyticsConnectorProfileCredentials_oAuthRequest - The OAuth requirement needed to request security tokens from the connector endpoint.

$sel:clientId:GoogleAnalyticsConnectorProfileCredentials', googleAnalyticsConnectorProfileCredentials_clientId - The identifier for the desired client.

$sel:clientSecret:GoogleAnalyticsConnectorProfileCredentials', googleAnalyticsConnectorProfileCredentials_clientSecret - The client secret used by the OAuth client to authenticate to the authorization server.

googleAnalyticsConnectorProfileCredentials_accessToken :: Lens' GoogleAnalyticsConnectorProfileCredentials (Maybe Text)

The credentials used to access protected Google Analytics resources.

googleAnalyticsConnectorProfileCredentials_refreshToken :: Lens' GoogleAnalyticsConnectorProfileCredentials (Maybe Text)

The credentials used to acquire new access tokens. This is required only for OAuth2 access tokens, and is not required for OAuth1 access tokens.

googleAnalyticsConnectorProfileCredentials_oAuthRequest :: Lens' GoogleAnalyticsConnectorProfileCredentials (Maybe ConnectorOAuthRequest)

The OAuth requirement needed to request security tokens from the connector endpoint.

googleAnalyticsConnectorProfileCredentials_clientId :: Lens' GoogleAnalyticsConnectorProfileCredentials Text

The identifier for the desired client.

googleAnalyticsConnectorProfileCredentials_clientSecret :: Lens' GoogleAnalyticsConnectorProfileCredentials Text

The client secret used by the OAuth client to authenticate to the authorization server.