Amazonka.AppFlow.Types.SalesforceConnectorProfileCredentials

Description

Documentation

data SalesforceConnectorProfileCredentials

The connector-specific profile credentials required when using Salesforce.

See: newSalesforceConnectorProfileCredentials smart constructor.

Constructors

SalesforceConnectorProfileCredentials'

Fields

accessToken :: Maybe (Sensitive Text)

The credentials used to access protected Salesforce resources.

refreshToken :: Maybe Text

The credentials used to acquire new access tokens.

oAuthRequest :: Maybe ConnectorOAuthRequest

The OAuth requirement needed to request security tokens from the connector endpoint.

clientCredentialsArn :: Maybe (Sensitive Text)

The secret manager ARN, which contains the client ID and client secret of the connected app.

newSalesforceConnectorProfileCredentials :: SalesforceConnectorProfileCredentials

Create a value of SalesforceConnectorProfileCredentials with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:accessToken:SalesforceConnectorProfileCredentials', salesforceConnectorProfileCredentials_accessToken - The credentials used to access protected Salesforce resources.

$sel:refreshToken:SalesforceConnectorProfileCredentials', salesforceConnectorProfileCredentials_refreshToken - The credentials used to acquire new access tokens.

$sel:oAuthRequest:SalesforceConnectorProfileCredentials', salesforceConnectorProfileCredentials_oAuthRequest - The OAuth requirement needed to request security tokens from the connector endpoint.

$sel:clientCredentialsArn:SalesforceConnectorProfileCredentials', salesforceConnectorProfileCredentials_clientCredentialsArn - The secret manager ARN, which contains the client ID and client secret of the connected app.

salesforceConnectorProfileCredentials_accessToken :: Lens' SalesforceConnectorProfileCredentials (Maybe Text)

The credentials used to access protected Salesforce resources.

salesforceConnectorProfileCredentials_refreshToken :: Lens' SalesforceConnectorProfileCredentials (Maybe Text)

The credentials used to acquire new access tokens.

salesforceConnectorProfileCredentials_oAuthRequest :: Lens' SalesforceConnectorProfileCredentials (Maybe ConnectorOAuthRequest)

The OAuth requirement needed to request security tokens from the connector endpoint.

salesforceConnectorProfileCredentials_clientCredentialsArn :: Lens' SalesforceConnectorProfileCredentials (Maybe Text)

The secret manager ARN, which contains the client ID and client secret of the connected app.