Amazonka.AppFlow.Types.SnowflakeConnectorProfileProperties

Description

Documentation

data SnowflakeConnectorProfileProperties

The connector-specific profile properties required when using Snowflake.

See: newSnowflakeConnectorProfileProperties smart constructor.

Constructors

SnowflakeConnectorProfileProperties'

Fields

privateLinkServiceName :: Maybe Text

The Snowflake Private Link service name to be used for private data transfers.

accountName :: Maybe Text

The name of the account.

bucketPrefix :: Maybe Text

The bucket path that refers to the Amazon S3 bucket associated with Snowflake.

region :: Maybe Text

The Amazon Web Services Region of the Snowflake account.

warehouse :: Text

The name of the Snowflake warehouse.

stage :: Text

The name of the Amazon S3 stage that was created while setting up an Amazon S3 stage in the Snowflake account. This is written in the following format: < Database>< Schema><Stage Name>.

bucketName :: Text

The name of the Amazon S3 bucket associated with Snowflake.

newSnowflakeConnectorProfileProperties

Arguments

:: Text
$sel:warehouse:SnowflakeConnectorProfileProperties'

-> Text
$sel:stage:SnowflakeConnectorProfileProperties'

-> Text
$sel:bucketName:SnowflakeConnectorProfileProperties'

-> SnowflakeConnectorProfileProperties

Create a value of SnowflakeConnectorProfileProperties with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:privateLinkServiceName:SnowflakeConnectorProfileProperties', snowflakeConnectorProfileProperties_privateLinkServiceName - The Snowflake Private Link service name to be used for private data transfers.

$sel:accountName:SnowflakeConnectorProfileProperties', snowflakeConnectorProfileProperties_accountName - The name of the account.

$sel:bucketPrefix:SnowflakeConnectorProfileProperties', snowflakeConnectorProfileProperties_bucketPrefix - The bucket path that refers to the Amazon S3 bucket associated with Snowflake.

$sel:region:SnowflakeConnectorProfileProperties', snowflakeConnectorProfileProperties_region - The Amazon Web Services Region of the Snowflake account.

$sel:warehouse:SnowflakeConnectorProfileProperties', snowflakeConnectorProfileProperties_warehouse - The name of the Snowflake warehouse.

$sel:stage:SnowflakeConnectorProfileProperties', snowflakeConnectorProfileProperties_stage - The name of the Amazon S3 stage that was created while setting up an Amazon S3 stage in the Snowflake account. This is written in the following format: < Database>< Schema><Stage Name>.

$sel:bucketName:SnowflakeConnectorProfileProperties', snowflakeConnectorProfileProperties_bucketName - The name of the Amazon S3 bucket associated with Snowflake.

snowflakeConnectorProfileProperties_privateLinkServiceName :: Lens' SnowflakeConnectorProfileProperties (Maybe Text)

The Snowflake Private Link service name to be used for private data transfers.

snowflakeConnectorProfileProperties_accountName :: Lens' SnowflakeConnectorProfileProperties (Maybe Text)

The name of the account.

snowflakeConnectorProfileProperties_bucketPrefix :: Lens' SnowflakeConnectorProfileProperties (Maybe Text)

The bucket path that refers to the Amazon S3 bucket associated with Snowflake.

snowflakeConnectorProfileProperties_region :: Lens' SnowflakeConnectorProfileProperties (Maybe Text)

The Amazon Web Services Region of the Snowflake account.

snowflakeConnectorProfileProperties_warehouse :: Lens' SnowflakeConnectorProfileProperties Text

The name of the Snowflake warehouse.

snowflakeConnectorProfileProperties_stage :: Lens' SnowflakeConnectorProfileProperties Text

The name of the Amazon S3 stage that was created while setting up an Amazon S3 stage in the Snowflake account. This is written in the following format: < Database>< Schema><Stage Name>.

snowflakeConnectorProfileProperties_bucketName :: Lens' SnowflakeConnectorProfileProperties Text

The name of the Amazon S3 bucket associated with Snowflake.