Amazonka.AppFlow.Types.Task

Description

Documentation

data Task

A class for modeling different type of tasks. Task implementation varies based on the TaskType.

See: newTask smart constructor.

Constructors

Task'

Fields

taskProperties :: Maybe (HashMap OperatorPropertiesKeys Text)

A map used to store task-related information. The execution service looks for particular information based on the TaskType.

connectorOperator :: Maybe ConnectorOperator

The operation to be performed on the provided source fields.

destinationField :: Maybe Text

A field in a destination connector, or a field value against which Amazon AppFlow validates a source field.

sourceFields :: [Text]

The source fields to which a particular task is applied.

taskType :: TaskType

Specifies the particular task implementation that Amazon AppFlow performs.

newTask

Arguments

:: TaskType
$sel:taskType:Task'

-> Task

Create a value of Task with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:taskProperties:Task', task_taskProperties - A map used to store task-related information. The execution service looks for particular information based on the TaskType.

$sel:connectorOperator:Task', task_connectorOperator - The operation to be performed on the provided source fields.

$sel:destinationField:Task', task_destinationField - A field in a destination connector, or a field value against which Amazon AppFlow validates a source field.

$sel:sourceFields:Task', task_sourceFields - The source fields to which a particular task is applied.

$sel:taskType:Task', task_taskType - Specifies the particular task implementation that Amazon AppFlow performs.

task_taskProperties :: Lens' Task (Maybe (HashMap OperatorPropertiesKeys Text))

A map used to store task-related information. The execution service looks for particular information based on the TaskType.

task_connectorOperator :: Lens' Task (Maybe ConnectorOperator)

The operation to be performed on the provided source fields.

task_destinationField :: Lens' Task (Maybe Text)

A field in a destination connector, or a field value against which Amazon AppFlow validates a source field.

task_sourceFields :: Lens' Task [Text]

The source fields to which a particular task is applied.

task_taskType :: Lens' Task TaskType

Specifies the particular task implementation that Amazon AppFlow performs.