Amazonka.AppFlow.Types.ZendeskConnectorProfileCredentials

Description

Documentation

data ZendeskConnectorProfileCredentials

The connector-specific profile credentials required when using Zendesk.

See: newZendeskConnectorProfileCredentials smart constructor.

Constructors

ZendeskConnectorProfileCredentials'

Fields

accessToken :: Maybe (Sensitive Text)

The credentials used to access protected Zendesk resources.

oAuthRequest :: Maybe ConnectorOAuthRequest

The OAuth requirement needed to request security tokens from the connector endpoint.

clientId :: Text

The identifier for the desired client.

clientSecret :: Sensitive Text

The client secret used by the OAuth client to authenticate to the authorization server.

newZendeskConnectorProfileCredentials

Arguments

:: Text
$sel:clientId:ZendeskConnectorProfileCredentials'

-> Text
$sel:clientSecret:ZendeskConnectorProfileCredentials'

-> ZendeskConnectorProfileCredentials

Create a value of ZendeskConnectorProfileCredentials with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:accessToken:ZendeskConnectorProfileCredentials', zendeskConnectorProfileCredentials_accessToken - The credentials used to access protected Zendesk resources.

$sel:oAuthRequest:ZendeskConnectorProfileCredentials', zendeskConnectorProfileCredentials_oAuthRequest - The OAuth requirement needed to request security tokens from the connector endpoint.

$sel:clientId:ZendeskConnectorProfileCredentials', zendeskConnectorProfileCredentials_clientId - The identifier for the desired client.

$sel:clientSecret:ZendeskConnectorProfileCredentials', zendeskConnectorProfileCredentials_clientSecret - The client secret used by the OAuth client to authenticate to the authorization server.

zendeskConnectorProfileCredentials_accessToken :: Lens' ZendeskConnectorProfileCredentials (Maybe Text)

The credentials used to access protected Zendesk resources.

zendeskConnectorProfileCredentials_oAuthRequest :: Lens' ZendeskConnectorProfileCredentials (Maybe ConnectorOAuthRequest)

The OAuth requirement needed to request security tokens from the connector endpoint.

zendeskConnectorProfileCredentials_clientId :: Lens' ZendeskConnectorProfileCredentials Text

The identifier for the desired client.

zendeskConnectorProfileCredentials_clientSecret :: Lens' ZendeskConnectorProfileCredentials Text

The client secret used by the OAuth client to authenticate to the authorization server.