Amazonka.AppFlow.UpdateConnectorProfile

Description

Updates a given connector profile associated with your account.

Creating a Request

data UpdateConnectorProfile

See: newUpdateConnectorProfile smart constructor.

Constructors

UpdateConnectorProfile'

Fields

connectorProfileName :: Text

The name of the connector profile and is unique for each ConnectorProfile in the Amazon Web Services account.

connectionMode :: ConnectionMode

Indicates the connection mode and if it is public or private.

connectorProfileConfig :: ConnectorProfileConfig

Defines the connector-specific profile configuration and credentials.

newUpdateConnectorProfile

Arguments

:: Text
$sel:connectorProfileName:UpdateConnectorProfile'

-> ConnectionMode
$sel:connectionMode:UpdateConnectorProfile'

-> ConnectorProfileConfig
$sel:connectorProfileConfig:UpdateConnectorProfile'

-> UpdateConnectorProfile

Create a value of UpdateConnectorProfile with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:connectorProfileName:UpdateConnectorProfile', updateConnectorProfile_connectorProfileName - The name of the connector profile and is unique for each ConnectorProfile in the Amazon Web Services account.

$sel:connectionMode:UpdateConnectorProfile', updateConnectorProfile_connectionMode - Indicates the connection mode and if it is public or private.

$sel:connectorProfileConfig:UpdateConnectorProfile', updateConnectorProfile_connectorProfileConfig - Defines the connector-specific profile configuration and credentials.

Request Lenses

updateConnectorProfile_connectorProfileName :: Lens' UpdateConnectorProfile Text

The name of the connector profile and is unique for each ConnectorProfile in the Amazon Web Services account.

updateConnectorProfile_connectionMode :: Lens' UpdateConnectorProfile ConnectionMode

Indicates the connection mode and if it is public or private.

updateConnectorProfile_connectorProfileConfig :: Lens' UpdateConnectorProfile ConnectorProfileConfig

Defines the connector-specific profile configuration and credentials.

Destructuring the Response

data UpdateConnectorProfileResponse

See: newUpdateConnectorProfileResponse smart constructor.

Constructors

UpdateConnectorProfileResponse'

Fields

connectorProfileArn :: Maybe Text

The Amazon Resource Name (ARN) of the connector profile.

httpStatus :: Int

The response's http status code.

newUpdateConnectorProfileResponse

Arguments

:: Int
$sel:httpStatus:UpdateConnectorProfileResponse'

-> UpdateConnectorProfileResponse

Create a value of UpdateConnectorProfileResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:connectorProfileArn:UpdateConnectorProfileResponse', updateConnectorProfileResponse_connectorProfileArn - The Amazon Resource Name (ARN) of the connector profile.

$sel:httpStatus:UpdateConnectorProfileResponse', updateConnectorProfileResponse_httpStatus - The response's http status code.

Response Lenses

updateConnectorProfileResponse_connectorProfileArn :: Lens' UpdateConnectorProfileResponse (Maybe Text)

The Amazon Resource Name (ARN) of the connector profile.

updateConnectorProfileResponse_httpStatus :: Lens' UpdateConnectorProfileResponse Int

The response's http status code.