Amazonka.AppIntegrationS

Description

Derived from API version 2020-07-29 of the AWS service descriptions, licensed under Apache 2.0.

The Amazon AppIntegrations service enables you to configure and reuse connections to external applications.

For information about how you can use external applications with Amazon Connect, see Set up pre-built integrations and Deliver information to agents using Amazon Connect Wisdom in the Amazon Connect Administrator Guide.

Service Configuration

defaultService :: Service

API version 2020-07-29 of the Amazon AppIntegrations Service SDK configuration.

Errors

Error matchers are designed for use with the functions provided by Control.Exception.Lens. This allows catching (and rethrowing) service specific errors returned by AppIntegrationS.

AccessDeniedException

_AccessDeniedException :: AsError a => Getting (First ServiceError) a ServiceError

You do not have sufficient access to perform this action.

InvalidRequestException

_InvalidRequestException :: AsError a => Getting (First ServiceError) a ServiceError

The request is not valid.

DuplicateResourceException

_DuplicateResourceException :: AsError a => Getting (First ServiceError) a ServiceError

A resource with the specified name already exists.

ResourceQuotaExceededException

_ResourceQuotaExceededException :: AsError a => Getting (First ServiceError) a ServiceError

The allowed quota for the resource has been exceeded.

ThrottlingException

_ThrottlingException :: AsError a => Getting (First ServiceError) a ServiceError

The throttling limit has been exceeded.

InternalServiceError

_InternalServiceError :: AsError a => Getting (First ServiceError) a ServiceError

Request processing failed due to an error or failure with the service.

ResourceNotFoundException

_ResourceNotFoundException :: AsError a => Getting (First ServiceError) a ServiceError

The specified resource was not found.

Waiters

Waiters poll by repeatedly sending a request until some remote success condition configured by the Wait specification is fulfilled. The Wait specification determines how many attempts should be made, in addition to delay and retry strategies.

Operations

Some AWS operations return results that are incomplete and require subsequent requests in order to obtain the entire result set. The process of sending subsequent requests to continue where a previous request left off is called pagination. For example, the ListObjects operation of Amazon S3 returns up to 1000 objects at a time, and you must send subsequent requests with the appropriate Marker in order to retrieve the next page of results.

Operations that have an AWSPager instance can transparently perform subsequent requests, correctly setting Markers and other request facets to iterate through the entire result set of a truncated API operation. Operations which support this have an additional note in the documentation.

Many operations have the ability to filter results on the server side. See the individual operation parameters for details.

GetEventIntegration

data GetEventIntegration

See: newGetEventIntegration smart constructor.

Constructors

GetEventIntegration' Text

newGetEventIntegration

Arguments

:: Text
$sel:name:GetEventIntegration'

-> GetEventIntegration

Create a value of GetEventIntegration with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:name:GetEventIntegration', getEventIntegration_name - The name of the event integration.

data GetEventIntegrationResponse

See: newGetEventIntegrationResponse smart constructor.

Constructors

GetEventIntegrationResponse' (Maybe Text) (Maybe EventFilter) (Maybe Text) (Maybe Text) (Maybe Text) (Maybe (HashMap Text Text)) Int

newGetEventIntegrationResponse

Arguments

:: Int
$sel:httpStatus:GetEventIntegrationResponse'

-> GetEventIntegrationResponse

Create a value of GetEventIntegrationResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:eventBridgeBus:GetEventIntegrationResponse', getEventIntegrationResponse_eventBridgeBus - The EventBridge bus.

$sel:eventFilter:GetEventIntegrationResponse', getEventIntegrationResponse_eventFilter - The event filter.

$sel:eventIntegrationArn:GetEventIntegrationResponse', getEventIntegrationResponse_eventIntegrationArn - The Amazon Resource Name (ARN) for the event integration.

$sel:name:GetEventIntegration', getEventIntegrationResponse_name - The name of the event integration.

$sel:description:GetEventIntegrationResponse', getEventIntegrationResponse_description - The description of the event integration.

$sel:tags:GetEventIntegrationResponse', getEventIntegrationResponse_tags - One or more tags.

$sel:httpStatus:GetEventIntegrationResponse', getEventIntegrationResponse_httpStatus - The response's http status code.

ListDataIntegrations

data ListDataIntegrations

See: newListDataIntegrations smart constructor.

Constructors

ListDataIntegrations' (Maybe Text) (Maybe Natural)

newListDataIntegrations :: ListDataIntegrations

Create a value of ListDataIntegrations with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:nextToken:ListDataIntegrations', listDataIntegrations_nextToken - The token for the next set of results. Use the value returned in the previous response in the next request to retrieve the next set of results.

$sel:maxResults:ListDataIntegrations', listDataIntegrations_maxResults - The maximum number of results to return per page.

data ListDataIntegrationsResponse

See: newListDataIntegrationsResponse smart constructor.

Constructors

ListDataIntegrationsResponse' (Maybe (NonEmpty DataIntegrationSummary)) (Maybe Text) Int

newListDataIntegrationsResponse

Arguments

:: Int
$sel:httpStatus:ListDataIntegrationsResponse'

-> ListDataIntegrationsResponse

Create a value of ListDataIntegrationsResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:dataIntegrations:ListDataIntegrationsResponse', listDataIntegrationsResponse_dataIntegrations - The DataIntegrations associated with this account.

$sel:nextToken:ListDataIntegrations', listDataIntegrationsResponse_nextToken - If there are additional results, this is the token for the next set of results.

$sel:httpStatus:ListDataIntegrationsResponse', listDataIntegrationsResponse_httpStatus - The response's http status code.

CreateDataIntegration

data CreateDataIntegration

See: newCreateDataIntegration smart constructor.

Constructors

CreateDataIntegration' (Maybe ScheduleConfiguration) (Maybe Text) (Maybe Text) (Maybe Text) (Maybe Text) (Maybe (HashMap Text Text)) Text

newCreateDataIntegration

Arguments

:: Text
$sel:name:CreateDataIntegration'

-> CreateDataIntegration

Create a value of CreateDataIntegration with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:scheduleConfig:CreateDataIntegration', createDataIntegration_scheduleConfig - The name of the data and how often it should be pulled from the source.

$sel:clientToken:CreateDataIntegration', createDataIntegration_clientToken - A unique, case-sensitive identifier that you provide to ensure the idempotency of the request.

$sel:kmsKey:CreateDataIntegration', createDataIntegration_kmsKey - The KMS key for the DataIntegration.

$sel:sourceURI:CreateDataIntegration', createDataIntegration_sourceURI - The URI of the data source.

$sel:description:CreateDataIntegration', createDataIntegration_description - A description of the DataIntegration.

$sel:tags:CreateDataIntegration', createDataIntegration_tags - One or more tags.

$sel:name:CreateDataIntegration', createDataIntegration_name - The name of the DataIntegration.

data CreateDataIntegrationResponse

See: newCreateDataIntegrationResponse smart constructor.

Constructors

CreateDataIntegrationResponse' (Maybe Text) (Maybe Text) (Maybe Text) (Maybe ScheduleConfiguration) (Maybe Text) (Maybe Text) (Maybe Text) (Maybe Text) (Maybe (HashMap Text Text)) Int

newCreateDataIntegrationResponse

Arguments

:: Int
$sel:httpStatus:CreateDataIntegrationResponse'

-> CreateDataIntegrationResponse

Create a value of CreateDataIntegrationResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:clientToken:CreateDataIntegration', createDataIntegrationResponse_clientToken - A unique, case-sensitive identifier that you provide to ensure the idempotency of the request.

$sel:kmsKey:CreateDataIntegration', createDataIntegrationResponse_kmsKey - The KMS key for the DataIntegration.

$sel:arn:CreateDataIntegrationResponse', createDataIntegrationResponse_arn - The Amazon Resource Name (ARN)

$sel:scheduleConfiguration:CreateDataIntegrationResponse', createDataIntegrationResponse_scheduleConfiguration - The name of the data and how often it should be pulled from the source.

$sel:name:CreateDataIntegration', createDataIntegrationResponse_name - The name of the DataIntegration.

$sel:id:CreateDataIntegrationResponse', createDataIntegrationResponse_id - A unique identifier.

$sel:sourceURI:CreateDataIntegration', createDataIntegrationResponse_sourceURI - The URI of the data source.

$sel:description:CreateDataIntegration', createDataIntegrationResponse_description - A description of the DataIntegration.

$sel:tags:CreateDataIntegration', createDataIntegrationResponse_tags - One or more tags.

$sel:httpStatus:CreateDataIntegrationResponse', createDataIntegrationResponse_httpStatus - The response's http status code.

ListTagsForResource

data ListTagsForResource

See: newListTagsForResource smart constructor.

Constructors

ListTagsForResource' Text

newListTagsForResource

Arguments

:: Text
$sel:resourceArn:ListTagsForResource'

-> ListTagsForResource

Create a value of ListTagsForResource with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:resourceArn:ListTagsForResource', listTagsForResource_resourceArn - The Amazon Resource Name (ARN) of the resource.

data ListTagsForResourceResponse

See: newListTagsForResourceResponse smart constructor.

Constructors

ListTagsForResourceResponse' (Maybe (HashMap Text Text)) Int

newListTagsForResourceResponse

Arguments

:: Int
$sel:httpStatus:ListTagsForResourceResponse'

-> ListTagsForResourceResponse

Create a value of ListTagsForResourceResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:tags:ListTagsForResourceResponse', listTagsForResourceResponse_tags - Information about the tags.

$sel:httpStatus:ListTagsForResourceResponse', listTagsForResourceResponse_httpStatus - The response's http status code.

ListEventIntegrationAssociations

data ListEventIntegrationAssociations

See: newListEventIntegrationAssociations smart constructor.

Constructors

ListEventIntegrationAssociations' (Maybe Text) (Maybe Natural) Text

newListEventIntegrationAssociations

Arguments

:: Text
$sel:eventIntegrationName:ListEventIntegrationAssociations'

-> ListEventIntegrationAssociations

Create a value of ListEventIntegrationAssociations with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:nextToken:ListEventIntegrationAssociations', listEventIntegrationAssociations_nextToken - The token for the next set of results. Use the value returned in the previous response in the next request to retrieve the next set of results.

$sel:maxResults:ListEventIntegrationAssociations', listEventIntegrationAssociations_maxResults - The maximum number of results to return per page.

$sel:eventIntegrationName:ListEventIntegrationAssociations', listEventIntegrationAssociations_eventIntegrationName - The name of the event integration.

data ListEventIntegrationAssociationsResponse

See: newListEventIntegrationAssociationsResponse smart constructor.

Constructors

ListEventIntegrationAssociationsResponse' (Maybe (NonEmpty EventIntegrationAssociation)) (Maybe Text) Int

newListEventIntegrationAssociationsResponse

Arguments

:: Int
$sel:httpStatus:ListEventIntegrationAssociationsResponse'

-> ListEventIntegrationAssociationsResponse

Create a value of ListEventIntegrationAssociationsResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:eventIntegrationAssociations:ListEventIntegrationAssociationsResponse', listEventIntegrationAssociationsResponse_eventIntegrationAssociations - The event integration associations.

$sel:nextToken:ListEventIntegrationAssociations', listEventIntegrationAssociationsResponse_nextToken - If there are additional results, this is the token for the next set of results.

$sel:httpStatus:ListEventIntegrationAssociationsResponse', listEventIntegrationAssociationsResponse_httpStatus - The response's http status code.

GetDataIntegration

data GetDataIntegration

See: newGetDataIntegration smart constructor.

Constructors

GetDataIntegration' Text

newGetDataIntegration

Arguments

:: Text
$sel:identifier:GetDataIntegration'

-> GetDataIntegration

Create a value of GetDataIntegration with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:identifier:GetDataIntegration', getDataIntegration_identifier - A unique identifier.

data GetDataIntegrationResponse

See: newGetDataIntegrationResponse smart constructor.

Constructors

GetDataIntegrationResponse' (Maybe Text) (Maybe Text) (Maybe ScheduleConfiguration) (Maybe Text) (Maybe Text) (Maybe Text) (Maybe Text) (Maybe (HashMap Text Text)) Int

newGetDataIntegrationResponse

Arguments

:: Int
$sel:httpStatus:GetDataIntegrationResponse'

-> GetDataIntegrationResponse

Create a value of GetDataIntegrationResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:kmsKey:GetDataIntegrationResponse', getDataIntegrationResponse_kmsKey - The KMS key for the DataIntegration.

$sel:arn:GetDataIntegrationResponse', getDataIntegrationResponse_arn - The Amazon Resource Name (ARN) for the DataIntegration.

$sel:scheduleConfiguration:GetDataIntegrationResponse', getDataIntegrationResponse_scheduleConfiguration - The name of the data and how often it should be pulled from the source.

$sel:name:GetDataIntegrationResponse', getDataIntegrationResponse_name - The name of the DataIntegration.

$sel:id:GetDataIntegrationResponse', getDataIntegrationResponse_id - A unique identifier.

$sel:sourceURI:GetDataIntegrationResponse', getDataIntegrationResponse_sourceURI - The URI of the data source.

$sel:description:GetDataIntegrationResponse', getDataIntegrationResponse_description - The KMS key for the DataIntegration.

$sel:tags:GetDataIntegrationResponse', getDataIntegrationResponse_tags - One or more tags.

$sel:httpStatus:GetDataIntegrationResponse', getDataIntegrationResponse_httpStatus - The response's http status code.

ListEventIntegrations

data ListEventIntegrations

See: newListEventIntegrations smart constructor.

Constructors

ListEventIntegrations' (Maybe Text) (Maybe Natural)

newListEventIntegrations :: ListEventIntegrations

Create a value of ListEventIntegrations with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:nextToken:ListEventIntegrations', listEventIntegrations_nextToken - The token for the next set of results. Use the value returned in the previous response in the next request to retrieve the next set of results.

$sel:maxResults:ListEventIntegrations', listEventIntegrations_maxResults - The maximum number of results to return per page.

data ListEventIntegrationsResponse

See: newListEventIntegrationsResponse smart constructor.

Constructors

ListEventIntegrationsResponse' (Maybe (NonEmpty EventIntegration)) (Maybe Text) Int

newListEventIntegrationsResponse

Arguments

:: Int
$sel:httpStatus:ListEventIntegrationsResponse'

-> ListEventIntegrationsResponse

Create a value of ListEventIntegrationsResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:eventIntegrations:ListEventIntegrationsResponse', listEventIntegrationsResponse_eventIntegrations - The event integrations.

$sel:nextToken:ListEventIntegrations', listEventIntegrationsResponse_nextToken - If there are additional results, this is the token for the next set of results.

$sel:httpStatus:ListEventIntegrationsResponse', listEventIntegrationsResponse_httpStatus - The response's http status code.

DeleteEventIntegration

data DeleteEventIntegration

See: newDeleteEventIntegration smart constructor.

Constructors

DeleteEventIntegration' Text

newDeleteEventIntegration

Arguments

:: Text
$sel:name:DeleteEventIntegration'

-> DeleteEventIntegration

Create a value of DeleteEventIntegration with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:name:DeleteEventIntegration', deleteEventIntegration_name - The name of the event integration.

data DeleteEventIntegrationResponse

See: newDeleteEventIntegrationResponse smart constructor.

Constructors

DeleteEventIntegrationResponse' Int

newDeleteEventIntegrationResponse

Arguments

:: Int
$sel:httpStatus:DeleteEventIntegrationResponse'

-> DeleteEventIntegrationResponse

Create a value of DeleteEventIntegrationResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:httpStatus:DeleteEventIntegrationResponse', deleteEventIntegrationResponse_httpStatus - The response's http status code.

UpdateEventIntegration

data UpdateEventIntegration

See: newUpdateEventIntegration smart constructor.

Constructors

UpdateEventIntegration' (Maybe Text) Text

newUpdateEventIntegration

Arguments

:: Text
$sel:name:UpdateEventIntegration'

-> UpdateEventIntegration

Create a value of UpdateEventIntegration with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:description:UpdateEventIntegration', updateEventIntegration_description - The description of the event inegration.

$sel:name:UpdateEventIntegration', updateEventIntegration_name - The name of the event integration.

data UpdateEventIntegrationResponse

See: newUpdateEventIntegrationResponse smart constructor.

Constructors

UpdateEventIntegrationResponse' Int

newUpdateEventIntegrationResponse

Arguments

:: Int
$sel:httpStatus:UpdateEventIntegrationResponse'

-> UpdateEventIntegrationResponse

Create a value of UpdateEventIntegrationResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:httpStatus:UpdateEventIntegrationResponse', updateEventIntegrationResponse_httpStatus - The response's http status code.

DeleteDataIntegration

data DeleteDataIntegration

See: newDeleteDataIntegration smart constructor.

Constructors

DeleteDataIntegration' Text

newDeleteDataIntegration

Arguments

:: Text
$sel:dataIntegrationIdentifier:DeleteDataIntegration'

-> DeleteDataIntegration

Create a value of DeleteDataIntegration with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:dataIntegrationIdentifier:DeleteDataIntegration', deleteDataIntegration_dataIntegrationIdentifier - A unique identifier for the DataIntegration.

data DeleteDataIntegrationResponse

See: newDeleteDataIntegrationResponse smart constructor.

Constructors

DeleteDataIntegrationResponse' Int

newDeleteDataIntegrationResponse

Arguments

:: Int
$sel:httpStatus:DeleteDataIntegrationResponse'

-> DeleteDataIntegrationResponse

Create a value of DeleteDataIntegrationResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:httpStatus:DeleteDataIntegrationResponse', deleteDataIntegrationResponse_httpStatus - The response's http status code.

UpdateDataIntegration

data UpdateDataIntegration

See: newUpdateDataIntegration smart constructor.

Constructors

UpdateDataIntegration' (Maybe Text) (Maybe Text) Text

newUpdateDataIntegration

Arguments

:: Text
$sel:identifier:UpdateDataIntegration'

-> UpdateDataIntegration

Create a value of UpdateDataIntegration with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:name:UpdateDataIntegration', updateDataIntegration_name - The name of the DataIntegration.

$sel:description:UpdateDataIntegration', updateDataIntegration_description - A description of the DataIntegration.

$sel:identifier:UpdateDataIntegration', updateDataIntegration_identifier - A unique identifier for the DataIntegration.

data UpdateDataIntegrationResponse

See: newUpdateDataIntegrationResponse smart constructor.

Constructors

UpdateDataIntegrationResponse' Int

newUpdateDataIntegrationResponse

Arguments

:: Int
$sel:httpStatus:UpdateDataIntegrationResponse'

-> UpdateDataIntegrationResponse

Create a value of UpdateDataIntegrationResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:httpStatus:UpdateDataIntegrationResponse', updateDataIntegrationResponse_httpStatus - The response's http status code.

TagResource

data TagResource

See: newTagResource smart constructor.

Constructors

TagResource' Text (HashMap Text Text)

newTagResource

Arguments

:: Text
$sel:resourceArn:TagResource'

-> TagResource

Create a value of TagResource with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:resourceArn:TagResource', tagResource_resourceArn - The Amazon Resource Name (ARN) of the resource.

$sel:tags:TagResource', tagResource_tags - One or more tags.

data TagResourceResponse

See: newTagResourceResponse smart constructor.

Constructors

TagResourceResponse' Int

newTagResourceResponse

Arguments

:: Int
$sel:httpStatus:TagResourceResponse'

-> TagResourceResponse

Create a value of TagResourceResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:httpStatus:TagResourceResponse', tagResourceResponse_httpStatus - The response's http status code.

UntagResource

data UntagResource

See: newUntagResource smart constructor.

Constructors

UntagResource' Text (NonEmpty Text)

newUntagResource

Arguments

:: Text
$sel:resourceArn:UntagResource'

-> NonEmpty Text
$sel:tagKeys:UntagResource'

-> UntagResource

Create a value of UntagResource with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:resourceArn:UntagResource', untagResource_resourceArn - The Amazon Resource Name (ARN) of the resource.

$sel:tagKeys:UntagResource', untagResource_tagKeys - The tag keys.

data UntagResourceResponse

See: newUntagResourceResponse smart constructor.

Constructors

UntagResourceResponse' Int

newUntagResourceResponse

Arguments

:: Int
$sel:httpStatus:UntagResourceResponse'

-> UntagResourceResponse

Create a value of UntagResourceResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:httpStatus:UntagResourceResponse', untagResourceResponse_httpStatus - The response's http status code.

ListDataIntegrationAssociations

data ListDataIntegrationAssociations

See: newListDataIntegrationAssociations smart constructor.

Constructors

ListDataIntegrationAssociations' (Maybe Text) (Maybe Natural) Text

newListDataIntegrationAssociations

Arguments

:: Text
$sel:dataIntegrationIdentifier:ListDataIntegrationAssociations'

-> ListDataIntegrationAssociations

Create a value of ListDataIntegrationAssociations with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:nextToken:ListDataIntegrationAssociations', listDataIntegrationAssociations_nextToken - The token for the next set of results. Use the value returned in the previous response in the next request to retrieve the next set of results.

$sel:maxResults:ListDataIntegrationAssociations', listDataIntegrationAssociations_maxResults - The maximum number of results to return per page.

$sel:dataIntegrationIdentifier:ListDataIntegrationAssociations', listDataIntegrationAssociations_dataIntegrationIdentifier - A unique identifier for the DataIntegration.

data ListDataIntegrationAssociationsResponse

See: newListDataIntegrationAssociationsResponse smart constructor.

Constructors

ListDataIntegrationAssociationsResponse' (Maybe (NonEmpty DataIntegrationAssociationSummary)) (Maybe Text) Int

newListDataIntegrationAssociationsResponse

Arguments

:: Int
$sel:httpStatus:ListDataIntegrationAssociationsResponse'

-> ListDataIntegrationAssociationsResponse

Create a value of ListDataIntegrationAssociationsResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:dataIntegrationAssociations:ListDataIntegrationAssociationsResponse', listDataIntegrationAssociationsResponse_dataIntegrationAssociations - The Amazon Resource Name (ARN) and unique ID of the DataIntegration association.

$sel:nextToken:ListDataIntegrationAssociations', listDataIntegrationAssociationsResponse_nextToken - If there are additional results, this is the token for the next set of results.

$sel:httpStatus:ListDataIntegrationAssociationsResponse', listDataIntegrationAssociationsResponse_httpStatus - The response's http status code.

CreateEventIntegration

data CreateEventIntegration

See: newCreateEventIntegration smart constructor.

Constructors

CreateEventIntegration' (Maybe Text) (Maybe Text) (Maybe (HashMap Text Text)) Text EventFilter Text

newCreateEventIntegration

Arguments

:: Text
$sel:name:CreateEventIntegration'

-> EventFilter
$sel:eventFilter:CreateEventIntegration'

-> Text
$sel:eventBridgeBus:CreateEventIntegration'

-> CreateEventIntegration

Create a value of CreateEventIntegration with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:clientToken:CreateEventIntegration', createEventIntegration_clientToken - A unique, case-sensitive identifier that you provide to ensure the idempotency of the request.

$sel:description:CreateEventIntegration', createEventIntegration_description - The description of the event integration.

$sel:tags:CreateEventIntegration', createEventIntegration_tags - One or more tags.

$sel:name:CreateEventIntegration', createEventIntegration_name - The name of the event integration.

$sel:eventFilter:CreateEventIntegration', createEventIntegration_eventFilter - The event filter.

$sel:eventBridgeBus:CreateEventIntegration', createEventIntegration_eventBridgeBus - The EventBridge bus.

data CreateEventIntegrationResponse

See: newCreateEventIntegrationResponse smart constructor.

Constructors

CreateEventIntegrationResponse' (Maybe Text) Int

newCreateEventIntegrationResponse

Arguments

:: Int
$sel:httpStatus:CreateEventIntegrationResponse'

-> CreateEventIntegrationResponse

Create a value of CreateEventIntegrationResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:eventIntegrationArn:CreateEventIntegrationResponse', createEventIntegrationResponse_eventIntegrationArn - The Amazon Resource Name (ARN) of the event integration.

$sel:httpStatus:CreateEventIntegrationResponse', createEventIntegrationResponse_httpStatus - The response's http status code.

Types

DataIntegrationAssociationSummary

data DataIntegrationAssociationSummary

Summary information about the DataIntegration association.

See: newDataIntegrationAssociationSummary smart constructor.

Constructors

DataIntegrationAssociationSummary' (Maybe Text) (Maybe Text) (Maybe Text)

newDataIntegrationAssociationSummary :: DataIntegrationAssociationSummary

Create a value of DataIntegrationAssociationSummary with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:clientId:DataIntegrationAssociationSummary', dataIntegrationAssociationSummary_clientId - The identifier for teh client that is associated with the DataIntegration association.

$sel:dataIntegrationAssociationArn:DataIntegrationAssociationSummary', dataIntegrationAssociationSummary_dataIntegrationAssociationArn - The Amazon Resource Name (ARN) of the DataIntegration association.

$sel:dataIntegrationArn:DataIntegrationAssociationSummary', dataIntegrationAssociationSummary_dataIntegrationArn - The Amazon Resource Name (ARN)of the DataIntegration.

DataIntegrationSummary

data DataIntegrationSummary

Summary information about the DataIntegration.

See: newDataIntegrationSummary smart constructor.

Constructors

DataIntegrationSummary' (Maybe Text) (Maybe Text) (Maybe Text)

newDataIntegrationSummary :: DataIntegrationSummary

Create a value of DataIntegrationSummary with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:arn:DataIntegrationSummary', dataIntegrationSummary_arn - The Amazon Resource Name (ARN) of the DataIntegration.

$sel:name:DataIntegrationSummary', dataIntegrationSummary_name - The name of the DataIntegration.

$sel:sourceURI:DataIntegrationSummary', dataIntegrationSummary_sourceURI - The URI of the data source.

EventFilter

data EventFilter

The event filter.

See: newEventFilter smart constructor.

Constructors

EventFilter' Text

newEventFilter

Arguments

:: Text
$sel:source:EventFilter'

-> EventFilter

Create a value of EventFilter with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:source:EventFilter', eventFilter_source - The source of the events.

EventIntegration

data EventIntegration

The event integration.

See: newEventIntegration smart constructor.

Constructors

EventIntegration' (Maybe Text) (Maybe EventFilter) (Maybe Text) (Maybe Text) (Maybe Text) (Maybe (HashMap Text Text))

newEventIntegration :: EventIntegration

Create a value of EventIntegration with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:eventBridgeBus:EventIntegration', eventIntegration_eventBridgeBus - The Amazon EventBridge bus for the event integration.

$sel:eventFilter:EventIntegration', eventIntegration_eventFilter - The event integration filter.

$sel:eventIntegrationArn:EventIntegration', eventIntegration_eventIntegrationArn - The Amazon Resource Name (ARN) of the event integration.

$sel:name:EventIntegration', eventIntegration_name - The name of the event integration.

$sel:description:EventIntegration', eventIntegration_description - The event integration description.

$sel:tags:EventIntegration', eventIntegration_tags - The tags.

EventIntegrationAssociation

data EventIntegrationAssociation

The event integration association.

See: newEventIntegrationAssociation smart constructor.

Constructors

EventIntegrationAssociation' (Maybe Text) (Maybe Text) (Maybe (HashMap Text Text)) (Maybe Text) (Maybe Text) (Maybe Text)

newEventIntegrationAssociation :: EventIntegrationAssociation

Create a value of EventIntegrationAssociation with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:clientId:EventIntegrationAssociation', eventIntegrationAssociation_clientId - The identifier for the client that is associated with the event integration.

$sel:eventIntegrationName:EventIntegrationAssociation', eventIntegrationAssociation_eventIntegrationName - The name of the event integration.

$sel:clientAssociationMetadata:EventIntegrationAssociation', eventIntegrationAssociation_clientAssociationMetadata - The metadata associated with the client.

$sel:eventIntegrationAssociationId:EventIntegrationAssociation', eventIntegrationAssociation_eventIntegrationAssociationId - The identifier for the event integration association.

$sel:eventIntegrationAssociationArn:EventIntegrationAssociation', eventIntegrationAssociation_eventIntegrationAssociationArn - The Amazon Resource Name (ARN) for the event integration association.

$sel:eventBridgeRuleName:EventIntegrationAssociation', eventIntegrationAssociation_eventBridgeRuleName - The name of the EventBridge rule.

ScheduleConfiguration

data ScheduleConfiguration

The name of the data and how often it should be pulled from the source.

See: newScheduleConfiguration smart constructor.

Constructors

ScheduleConfiguration' (Maybe Text) (Maybe Text) (Maybe Text)

newScheduleConfiguration :: ScheduleConfiguration

Create a value of ScheduleConfiguration with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:scheduleExpression:ScheduleConfiguration', scheduleConfiguration_scheduleExpression - How often the data should be pulled from data source.

$sel:object':ScheduleConfiguration', scheduleConfiguration_object - The name of the object to pull from the data source.

$sel:firstExecutionFrom:ScheduleConfiguration', scheduleConfiguration_firstExecutionFrom - The start date for objects to import in the first flow run.