Amazonka.AppIntegrationS.CreateEventIntegration

Description

Creates an EventIntegration, given a specified name, description, and a reference to an Amazon EventBridge bus in your account and a partner event source that pushes events to that bus. No objects are created in the your account, only metadata that is persisted on the EventIntegration control plane.

Creating a Request

data CreateEventIntegration

See: newCreateEventIntegration smart constructor.

Constructors

CreateEventIntegration'

Fields

clientToken :: Maybe Text

A unique, case-sensitive identifier that you provide to ensure the idempotency of the request.

description :: Maybe Text

The description of the event integration.

tags :: Maybe (HashMap Text Text)

One or more tags.

name :: Text

The name of the event integration.

eventFilter :: EventFilter

The event filter.

eventBridgeBus :: Text

The EventBridge bus.

newCreateEventIntegration

Arguments

:: Text
$sel:name:CreateEventIntegration'

-> EventFilter
$sel:eventFilter:CreateEventIntegration'

-> Text
$sel:eventBridgeBus:CreateEventIntegration'

-> CreateEventIntegration

Create a value of CreateEventIntegration with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:clientToken:CreateEventIntegration', createEventIntegration_clientToken - A unique, case-sensitive identifier that you provide to ensure the idempotency of the request.

$sel:description:CreateEventIntegration', createEventIntegration_description - The description of the event integration.

$sel:tags:CreateEventIntegration', createEventIntegration_tags - One or more tags.

$sel:name:CreateEventIntegration', createEventIntegration_name - The name of the event integration.

$sel:eventFilter:CreateEventIntegration', createEventIntegration_eventFilter - The event filter.

$sel:eventBridgeBus:CreateEventIntegration', createEventIntegration_eventBridgeBus - The EventBridge bus.

Request Lenses

createEventIntegration_clientToken :: Lens' CreateEventIntegration (Maybe Text)

A unique, case-sensitive identifier that you provide to ensure the idempotency of the request.

createEventIntegration_description :: Lens' CreateEventIntegration (Maybe Text)

The description of the event integration.

createEventIntegration_tags :: Lens' CreateEventIntegration (Maybe (HashMap Text Text))

One or more tags.

createEventIntegration_name :: Lens' CreateEventIntegration Text

The name of the event integration.

createEventIntegration_eventFilter :: Lens' CreateEventIntegration EventFilter

The event filter.

createEventIntegration_eventBridgeBus :: Lens' CreateEventIntegration Text

The EventBridge bus.

Destructuring the Response

data CreateEventIntegrationResponse

See: newCreateEventIntegrationResponse smart constructor.

Constructors

CreateEventIntegrationResponse'

Fields

eventIntegrationArn :: Maybe Text

The Amazon Resource Name (ARN) of the event integration.

httpStatus :: Int

The response's http status code.

newCreateEventIntegrationResponse

Arguments

:: Int
$sel:httpStatus:CreateEventIntegrationResponse'

-> CreateEventIntegrationResponse

Create a value of CreateEventIntegrationResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:eventIntegrationArn:CreateEventIntegrationResponse', createEventIntegrationResponse_eventIntegrationArn - The Amazon Resource Name (ARN) of the event integration.

$sel:httpStatus:CreateEventIntegrationResponse', createEventIntegrationResponse_httpStatus - The response's http status code.

Response Lenses

createEventIntegrationResponse_eventIntegrationArn :: Lens' CreateEventIntegrationResponse (Maybe Text)

The Amazon Resource Name (ARN) of the event integration.

createEventIntegrationResponse_httpStatus :: Lens' CreateEventIntegrationResponse Int

The response's http status code.