Amazonka.AppIntegrationS.GetEventIntegration

Description

Returns information about the event integration.

Creating a Request

data GetEventIntegration

See: newGetEventIntegration smart constructor.

Constructors

GetEventIntegration'

Fields

name :: Text

The name of the event integration.

newGetEventIntegration

Arguments

:: Text
$sel:name:GetEventIntegration'

-> GetEventIntegration

Create a value of GetEventIntegration with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:name:GetEventIntegration', getEventIntegration_name - The name of the event integration.

Request Lenses

getEventIntegration_name :: Lens' GetEventIntegration Text

The name of the event integration.

Destructuring the Response

data GetEventIntegrationResponse

See: newGetEventIntegrationResponse smart constructor.

Constructors

GetEventIntegrationResponse'

Fields

eventBridgeBus :: Maybe Text

The EventBridge bus.

eventFilter :: Maybe EventFilter

The event filter.

eventIntegrationArn :: Maybe Text

The Amazon Resource Name (ARN) for the event integration.

name :: Maybe Text

The name of the event integration.

description :: Maybe Text

The description of the event integration.

tags :: Maybe (HashMap Text Text)

One or more tags.

httpStatus :: Int

The response's http status code.

newGetEventIntegrationResponse

Arguments

:: Int
$sel:httpStatus:GetEventIntegrationResponse'

-> GetEventIntegrationResponse

Create a value of GetEventIntegrationResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:eventBridgeBus:GetEventIntegrationResponse', getEventIntegrationResponse_eventBridgeBus - The EventBridge bus.

$sel:eventFilter:GetEventIntegrationResponse', getEventIntegrationResponse_eventFilter - The event filter.

$sel:eventIntegrationArn:GetEventIntegrationResponse', getEventIntegrationResponse_eventIntegrationArn - The Amazon Resource Name (ARN) for the event integration.

$sel:name:GetEventIntegration', getEventIntegrationResponse_name - The name of the event integration.

$sel:description:GetEventIntegrationResponse', getEventIntegrationResponse_description - The description of the event integration.

$sel:tags:GetEventIntegrationResponse', getEventIntegrationResponse_tags - One or more tags.

$sel:httpStatus:GetEventIntegrationResponse', getEventIntegrationResponse_httpStatus - The response's http status code.

Response Lenses

getEventIntegrationResponse_eventBridgeBus :: Lens' GetEventIntegrationResponse (Maybe Text)

The EventBridge bus.

getEventIntegrationResponse_eventFilter :: Lens' GetEventIntegrationResponse (Maybe EventFilter)

The event filter.

getEventIntegrationResponse_eventIntegrationArn :: Lens' GetEventIntegrationResponse (Maybe Text)

The Amazon Resource Name (ARN) for the event integration.

getEventIntegrationResponse_name :: Lens' GetEventIntegrationResponse (Maybe Text)

The name of the event integration.

getEventIntegrationResponse_description :: Lens' GetEventIntegrationResponse (Maybe Text)

The description of the event integration.

getEventIntegrationResponse_tags :: Lens' GetEventIntegrationResponse (Maybe (HashMap Text Text))

One or more tags.

getEventIntegrationResponse_httpStatus :: Lens' GetEventIntegrationResponse Int

The response's http status code.