Amazonka.AppIntegrationS.Types

Description

Service Configuration

defaultService :: Service

API version 2020-07-29 of the Amazon AppIntegrations Service SDK configuration.

Errors

_AccessDeniedException :: AsError a => Getting (First ServiceError) a ServiceError

You do not have sufficient access to perform this action.

_InvalidRequestException :: AsError a => Getting (First ServiceError) a ServiceError

The request is not valid.

_DuplicateResourceException :: AsError a => Getting (First ServiceError) a ServiceError

A resource with the specified name already exists.

_ResourceQuotaExceededException :: AsError a => Getting (First ServiceError) a ServiceError

The allowed quota for the resource has been exceeded.

_ThrottlingException :: AsError a => Getting (First ServiceError) a ServiceError

The throttling limit has been exceeded.

_InternalServiceError :: AsError a => Getting (First ServiceError) a ServiceError

Request processing failed due to an error or failure with the service.

_ResourceNotFoundException :: AsError a => Getting (First ServiceError) a ServiceError

The specified resource was not found.

DataIntegrationAssociationSummary

data DataIntegrationAssociationSummary

Summary information about the DataIntegration association.

See: newDataIntegrationAssociationSummary smart constructor.

Constructors

DataIntegrationAssociationSummary'

Fields

clientId :: Maybe Text

The identifier for teh client that is associated with the DataIntegration association.

dataIntegrationAssociationArn :: Maybe Text

The Amazon Resource Name (ARN) of the DataIntegration association.

dataIntegrationArn :: Maybe Text

The Amazon Resource Name (ARN)of the DataIntegration.

newDataIntegrationAssociationSummary :: DataIntegrationAssociationSummary

Create a value of DataIntegrationAssociationSummary with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:clientId:DataIntegrationAssociationSummary', dataIntegrationAssociationSummary_clientId - The identifier for teh client that is associated with the DataIntegration association.

$sel:dataIntegrationAssociationArn:DataIntegrationAssociationSummary', dataIntegrationAssociationSummary_dataIntegrationAssociationArn - The Amazon Resource Name (ARN) of the DataIntegration association.

$sel:dataIntegrationArn:DataIntegrationAssociationSummary', dataIntegrationAssociationSummary_dataIntegrationArn - The Amazon Resource Name (ARN)of the DataIntegration.

dataIntegrationAssociationSummary_clientId :: Lens' DataIntegrationAssociationSummary (Maybe Text)

The identifier for teh client that is associated with the DataIntegration association.

dataIntegrationAssociationSummary_dataIntegrationAssociationArn :: Lens' DataIntegrationAssociationSummary (Maybe Text)

The Amazon Resource Name (ARN) of the DataIntegration association.

dataIntegrationAssociationSummary_dataIntegrationArn :: Lens' DataIntegrationAssociationSummary (Maybe Text)

The Amazon Resource Name (ARN)of the DataIntegration.

DataIntegrationSummary

data DataIntegrationSummary

Summary information about the DataIntegration.

See: newDataIntegrationSummary smart constructor.

Constructors

DataIntegrationSummary'

Fields

arn :: Maybe Text

The Amazon Resource Name (ARN) of the DataIntegration.

name :: Maybe Text

The name of the DataIntegration.

sourceURI :: Maybe Text

The URI of the data source.

newDataIntegrationSummary :: DataIntegrationSummary

Create a value of DataIntegrationSummary with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:arn:DataIntegrationSummary', dataIntegrationSummary_arn - The Amazon Resource Name (ARN) of the DataIntegration.

$sel:name:DataIntegrationSummary', dataIntegrationSummary_name - The name of the DataIntegration.

$sel:sourceURI:DataIntegrationSummary', dataIntegrationSummary_sourceURI - The URI of the data source.

dataIntegrationSummary_arn :: Lens' DataIntegrationSummary (Maybe Text)

The Amazon Resource Name (ARN) of the DataIntegration.

dataIntegrationSummary_name :: Lens' DataIntegrationSummary (Maybe Text)

The name of the DataIntegration.

dataIntegrationSummary_sourceURI :: Lens' DataIntegrationSummary (Maybe Text)

The URI of the data source.

EventFilter

data EventFilter

The event filter.

See: newEventFilter smart constructor.

Constructors

EventFilter'

Fields

source :: Text

The source of the events.

newEventFilter

Arguments

:: Text
$sel:source:EventFilter'

-> EventFilter

Create a value of EventFilter with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:source:EventFilter', eventFilter_source - The source of the events.

eventFilter_source :: Lens' EventFilter Text

The source of the events.

EventIntegration

data EventIntegration

The event integration.

See: newEventIntegration smart constructor.

Constructors

EventIntegration'

Fields

eventBridgeBus :: Maybe Text

The Amazon EventBridge bus for the event integration.

eventFilter :: Maybe EventFilter

The event integration filter.

eventIntegrationArn :: Maybe Text

The Amazon Resource Name (ARN) of the event integration.

name :: Maybe Text

The name of the event integration.

description :: Maybe Text

The event integration description.

tags :: Maybe (HashMap Text Text)

The tags.

newEventIntegration :: EventIntegration

Create a value of EventIntegration with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:eventBridgeBus:EventIntegration', eventIntegration_eventBridgeBus - The Amazon EventBridge bus for the event integration.

$sel:eventFilter:EventIntegration', eventIntegration_eventFilter - The event integration filter.

$sel:eventIntegrationArn:EventIntegration', eventIntegration_eventIntegrationArn - The Amazon Resource Name (ARN) of the event integration.

$sel:name:EventIntegration', eventIntegration_name - The name of the event integration.

$sel:description:EventIntegration', eventIntegration_description - The event integration description.

$sel:tags:EventIntegration', eventIntegration_tags - The tags.

eventIntegration_eventBridgeBus :: Lens' EventIntegration (Maybe Text)

The Amazon EventBridge bus for the event integration.

eventIntegration_eventFilter :: Lens' EventIntegration (Maybe EventFilter)

The event integration filter.

eventIntegration_eventIntegrationArn :: Lens' EventIntegration (Maybe Text)

The Amazon Resource Name (ARN) of the event integration.

eventIntegration_name :: Lens' EventIntegration (Maybe Text)

The name of the event integration.

eventIntegration_description :: Lens' EventIntegration (Maybe Text)

The event integration description.

eventIntegration_tags :: Lens' EventIntegration (Maybe (HashMap Text Text))

The tags.

EventIntegrationAssociation

data EventIntegrationAssociation

The event integration association.

See: newEventIntegrationAssociation smart constructor.

Constructors

EventIntegrationAssociation'

Fields

clientId :: Maybe Text

The identifier for the client that is associated with the event integration.

eventIntegrationName :: Maybe Text

The name of the event integration.

clientAssociationMetadata :: Maybe (HashMap Text Text)

The metadata associated with the client.

eventIntegrationAssociationId :: Maybe Text

The identifier for the event integration association.

eventIntegrationAssociationArn :: Maybe Text

The Amazon Resource Name (ARN) for the event integration association.

eventBridgeRuleName :: Maybe Text

The name of the EventBridge rule.

newEventIntegrationAssociation :: EventIntegrationAssociation

Create a value of EventIntegrationAssociation with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:clientId:EventIntegrationAssociation', eventIntegrationAssociation_clientId - The identifier for the client that is associated with the event integration.

$sel:eventIntegrationName:EventIntegrationAssociation', eventIntegrationAssociation_eventIntegrationName - The name of the event integration.

$sel:clientAssociationMetadata:EventIntegrationAssociation', eventIntegrationAssociation_clientAssociationMetadata - The metadata associated with the client.

$sel:eventIntegrationAssociationId:EventIntegrationAssociation', eventIntegrationAssociation_eventIntegrationAssociationId - The identifier for the event integration association.

$sel:eventIntegrationAssociationArn:EventIntegrationAssociation', eventIntegrationAssociation_eventIntegrationAssociationArn - The Amazon Resource Name (ARN) for the event integration association.

$sel:eventBridgeRuleName:EventIntegrationAssociation', eventIntegrationAssociation_eventBridgeRuleName - The name of the EventBridge rule.

eventIntegrationAssociation_clientId :: Lens' EventIntegrationAssociation (Maybe Text)

The identifier for the client that is associated with the event integration.

eventIntegrationAssociation_eventIntegrationName :: Lens' EventIntegrationAssociation (Maybe Text)

The name of the event integration.

eventIntegrationAssociation_clientAssociationMetadata :: Lens' EventIntegrationAssociation (Maybe (HashMap Text Text))

The metadata associated with the client.

eventIntegrationAssociation_eventIntegrationAssociationId :: Lens' EventIntegrationAssociation (Maybe Text)

The identifier for the event integration association.

eventIntegrationAssociation_eventIntegrationAssociationArn :: Lens' EventIntegrationAssociation (Maybe Text)

The Amazon Resource Name (ARN) for the event integration association.

eventIntegrationAssociation_eventBridgeRuleName :: Lens' EventIntegrationAssociation (Maybe Text)

The name of the EventBridge rule.

ScheduleConfiguration

data ScheduleConfiguration

The name of the data and how often it should be pulled from the source.

See: newScheduleConfiguration smart constructor.

Constructors

ScheduleConfiguration'

Fields

scheduleExpression :: Maybe Text

How often the data should be pulled from data source.

object' :: Maybe Text

The name of the object to pull from the data source.

firstExecutionFrom :: Maybe Text

The start date for objects to import in the first flow run.

newScheduleConfiguration :: ScheduleConfiguration

Create a value of ScheduleConfiguration with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:scheduleExpression:ScheduleConfiguration', scheduleConfiguration_scheduleExpression - How often the data should be pulled from data source.

$sel:object':ScheduleConfiguration', scheduleConfiguration_object - The name of the object to pull from the data source.

$sel:firstExecutionFrom:ScheduleConfiguration', scheduleConfiguration_firstExecutionFrom - The start date for objects to import in the first flow run.

scheduleConfiguration_scheduleExpression :: Lens' ScheduleConfiguration (Maybe Text)

How often the data should be pulled from data source.

scheduleConfiguration_object :: Lens' ScheduleConfiguration (Maybe Text)

The name of the object to pull from the data source.

scheduleConfiguration_firstExecutionFrom :: Lens' ScheduleConfiguration (Maybe Text)

The start date for objects to import in the first flow run.