Amazonka.AppMesh.CreateGatewayRoute

Description

Creates a gateway route.

A gateway route is attached to a virtual gateway and routes traffic to an existing virtual service. If a route matches a request, it can distribute traffic to a target virtual service.

For more information about gateway routes, see Gateway routes.

Creating a Request

data CreateGatewayRoute

See: newCreateGatewayRoute smart constructor.

Constructors

CreateGatewayRoute'

Fields

clientToken :: Maybe Text

Unique, case-sensitive identifier that you provide to ensure the idempotency of the request. Up to 36 letters, numbers, hyphens, and underscores are allowed.

meshOwner :: Maybe Text

The AWS IAM account ID of the service mesh owner. If the account ID is not your own, then the account that you specify must share the mesh with your account before you can create the resource in the service mesh. For more information about mesh sharing, see Working with shared meshes.

tags :: Maybe [TagRef]

Optional metadata that you can apply to the gateway route to assist with categorization and organization. Each tag consists of a key and an optional value, both of which you define. Tag keys can have a maximum character length of 128 characters, and tag values can have a maximum length of 256 characters.

gatewayRouteName :: Text

The name to use for the gateway route.

meshName :: Text

The name of the service mesh to create the gateway route in.

spec :: GatewayRouteSpec

The gateway route specification to apply.

virtualGatewayName :: Text

The name of the virtual gateway to associate the gateway route with. If the virtual gateway is in a shared mesh, then you must be the owner of the virtual gateway resource.

newCreateGatewayRoute

Arguments

:: Text
$sel:gatewayRouteName:CreateGatewayRoute'

-> Text
$sel:meshName:CreateGatewayRoute'

-> GatewayRouteSpec
$sel:spec:CreateGatewayRoute'

-> Text
$sel:virtualGatewayName:CreateGatewayRoute'

-> CreateGatewayRoute

Create a value of CreateGatewayRoute with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:clientToken:CreateGatewayRoute', createGatewayRoute_clientToken - Unique, case-sensitive identifier that you provide to ensure the idempotency of the request. Up to 36 letters, numbers, hyphens, and underscores are allowed.

$sel:meshOwner:CreateGatewayRoute', createGatewayRoute_meshOwner - The AWS IAM account ID of the service mesh owner. If the account ID is not your own, then the account that you specify must share the mesh with your account before you can create the resource in the service mesh. For more information about mesh sharing, see Working with shared meshes.

$sel:tags:CreateGatewayRoute', createGatewayRoute_tags - Optional metadata that you can apply to the gateway route to assist with categorization and organization. Each tag consists of a key and an optional value, both of which you define. Tag keys can have a maximum character length of 128 characters, and tag values can have a maximum length of 256 characters.

$sel:gatewayRouteName:CreateGatewayRoute', createGatewayRoute_gatewayRouteName - The name to use for the gateway route.

$sel:meshName:CreateGatewayRoute', createGatewayRoute_meshName - The name of the service mesh to create the gateway route in.

$sel:spec:CreateGatewayRoute', createGatewayRoute_spec - The gateway route specification to apply.

$sel:virtualGatewayName:CreateGatewayRoute', createGatewayRoute_virtualGatewayName - The name of the virtual gateway to associate the gateway route with. If the virtual gateway is in a shared mesh, then you must be the owner of the virtual gateway resource.

Request Lenses

createGatewayRoute_clientToken :: Lens' CreateGatewayRoute (Maybe Text)

Unique, case-sensitive identifier that you provide to ensure the idempotency of the request. Up to 36 letters, numbers, hyphens, and underscores are allowed.

createGatewayRoute_meshOwner :: Lens' CreateGatewayRoute (Maybe Text)

The AWS IAM account ID of the service mesh owner. If the account ID is not your own, then the account that you specify must share the mesh with your account before you can create the resource in the service mesh. For more information about mesh sharing, see Working with shared meshes.

createGatewayRoute_tags :: Lens' CreateGatewayRoute (Maybe [TagRef])

Optional metadata that you can apply to the gateway route to assist with categorization and organization. Each tag consists of a key and an optional value, both of which you define. Tag keys can have a maximum character length of 128 characters, and tag values can have a maximum length of 256 characters.

createGatewayRoute_gatewayRouteName :: Lens' CreateGatewayRoute Text

The name to use for the gateway route.

createGatewayRoute_meshName :: Lens' CreateGatewayRoute Text

The name of the service mesh to create the gateway route in.

createGatewayRoute_spec :: Lens' CreateGatewayRoute GatewayRouteSpec

The gateway route specification to apply.

createGatewayRoute_virtualGatewayName :: Lens' CreateGatewayRoute Text

The name of the virtual gateway to associate the gateway route with. If the virtual gateway is in a shared mesh, then you must be the owner of the virtual gateway resource.

Destructuring the Response

data CreateGatewayRouteResponse

See: newCreateGatewayRouteResponse smart constructor.

Constructors

CreateGatewayRouteResponse'

Fields

httpStatus :: Int

The response's http status code.

gatewayRoute :: GatewayRouteData

The full description of your gateway route following the create call.

newCreateGatewayRouteResponse

Arguments

:: Int
$sel:httpStatus:CreateGatewayRouteResponse'

-> GatewayRouteData
$sel:gatewayRoute:CreateGatewayRouteResponse'

-> CreateGatewayRouteResponse

Create a value of CreateGatewayRouteResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:httpStatus:CreateGatewayRouteResponse', createGatewayRouteResponse_httpStatus - The response's http status code.

$sel:gatewayRoute:CreateGatewayRouteResponse', createGatewayRouteResponse_gatewayRoute - The full description of your gateway route following the create call.

Response Lenses

createGatewayRouteResponse_httpStatus :: Lens' CreateGatewayRouteResponse Int

The response's http status code.

createGatewayRouteResponse_gatewayRoute :: Lens' CreateGatewayRouteResponse GatewayRouteData

The full description of your gateway route following the create call.