Amazonka.AppMesh.DeleteRoute

Description

Deletes an existing route.

Creating a Request

data DeleteRoute

See: newDeleteRoute smart constructor.

Constructors

DeleteRoute'

Fields

meshOwner :: Maybe Text

The AWS IAM account ID of the service mesh owner. If the account ID is not your own, then it's the ID of the account that shared the mesh with your account. For more information about mesh sharing, see Working with shared meshes.

meshName :: Text

The name of the service mesh to delete the route in.

routeName :: Text

The name of the route to delete.

virtualRouterName :: Text

The name of the virtual router to delete the route in.

newDeleteRoute

Arguments

:: Text
$sel:meshName:DeleteRoute'

-> Text
$sel:routeName:DeleteRoute'

-> Text
$sel:virtualRouterName:DeleteRoute'

-> DeleteRoute

Create a value of DeleteRoute with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:meshOwner:DeleteRoute', deleteRoute_meshOwner - The AWS IAM account ID of the service mesh owner. If the account ID is not your own, then it's the ID of the account that shared the mesh with your account. For more information about mesh sharing, see Working with shared meshes.

$sel:meshName:DeleteRoute', deleteRoute_meshName - The name of the service mesh to delete the route in.

$sel:routeName:DeleteRoute', deleteRoute_routeName - The name of the route to delete.

$sel:virtualRouterName:DeleteRoute', deleteRoute_virtualRouterName - The name of the virtual router to delete the route in.

Request Lenses

deleteRoute_meshOwner :: Lens' DeleteRoute (Maybe Text)

The AWS IAM account ID of the service mesh owner. If the account ID is not your own, then it's the ID of the account that shared the mesh with your account. For more information about mesh sharing, see Working with shared meshes.

deleteRoute_meshName :: Lens' DeleteRoute Text

The name of the service mesh to delete the route in.

deleteRoute_routeName :: Lens' DeleteRoute Text

The name of the route to delete.

deleteRoute_virtualRouterName :: Lens' DeleteRoute Text

The name of the virtual router to delete the route in.

Destructuring the Response

data DeleteRouteResponse

See: newDeleteRouteResponse smart constructor.

Constructors

DeleteRouteResponse'

Fields

httpStatus :: Int

The response's http status code.

route :: RouteData

The route that was deleted.

newDeleteRouteResponse

Arguments

:: Int
$sel:httpStatus:DeleteRouteResponse'

-> RouteData
$sel:route:DeleteRouteResponse'

-> DeleteRouteResponse

Create a value of DeleteRouteResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:httpStatus:DeleteRouteResponse', deleteRouteResponse_httpStatus - The response's http status code.

$sel:route:DeleteRouteResponse', deleteRouteResponse_route - The route that was deleted.

Response Lenses

deleteRouteResponse_httpStatus :: Lens' DeleteRouteResponse Int

The response's http status code.

deleteRouteResponse_route :: Lens' DeleteRouteResponse RouteData

The route that was deleted.