Amazonka.AppMesh.DeleteVirtualGateway

Description

Deletes an existing virtual gateway. You cannot delete a virtual gateway if any gateway routes are associated to it.

Creating a Request

data DeleteVirtualGateway

See: newDeleteVirtualGateway smart constructor.

Constructors

DeleteVirtualGateway'

Fields

meshOwner :: Maybe Text

The AWS IAM account ID of the service mesh owner. If the account ID is not your own, then it's the ID of the account that shared the mesh with your account. For more information about mesh sharing, see Working with shared meshes.

meshName :: Text

The name of the service mesh to delete the virtual gateway from.

virtualGatewayName :: Text

The name of the virtual gateway to delete.

newDeleteVirtualGateway

Arguments

:: Text
$sel:meshName:DeleteVirtualGateway'

-> Text
$sel:virtualGatewayName:DeleteVirtualGateway'

-> DeleteVirtualGateway

Create a value of DeleteVirtualGateway with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:meshOwner:DeleteVirtualGateway', deleteVirtualGateway_meshOwner - The AWS IAM account ID of the service mesh owner. If the account ID is not your own, then it's the ID of the account that shared the mesh with your account. For more information about mesh sharing, see Working with shared meshes.

$sel:meshName:DeleteVirtualGateway', deleteVirtualGateway_meshName - The name of the service mesh to delete the virtual gateway from.

$sel:virtualGatewayName:DeleteVirtualGateway', deleteVirtualGateway_virtualGatewayName - The name of the virtual gateway to delete.

Request Lenses

deleteVirtualGateway_meshOwner :: Lens' DeleteVirtualGateway (Maybe Text)

The AWS IAM account ID of the service mesh owner. If the account ID is not your own, then it's the ID of the account that shared the mesh with your account. For more information about mesh sharing, see Working with shared meshes.

deleteVirtualGateway_meshName :: Lens' DeleteVirtualGateway Text

The name of the service mesh to delete the virtual gateway from.

deleteVirtualGateway_virtualGatewayName :: Lens' DeleteVirtualGateway Text

The name of the virtual gateway to delete.

Destructuring the Response

data DeleteVirtualGatewayResponse

See: newDeleteVirtualGatewayResponse smart constructor.

Constructors

DeleteVirtualGatewayResponse'

Fields

httpStatus :: Int

The response's http status code.

virtualGateway :: VirtualGatewayData

The virtual gateway that was deleted.

newDeleteVirtualGatewayResponse

Arguments

:: Int
$sel:httpStatus:DeleteVirtualGatewayResponse'

-> VirtualGatewayData
$sel:virtualGateway:DeleteVirtualGatewayResponse'

-> DeleteVirtualGatewayResponse

Create a value of DeleteVirtualGatewayResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:httpStatus:DeleteVirtualGatewayResponse', deleteVirtualGatewayResponse_httpStatus - The response's http status code.

$sel:virtualGateway:DeleteVirtualGatewayResponse', deleteVirtualGatewayResponse_virtualGateway - The virtual gateway that was deleted.

Response Lenses

deleteVirtualGatewayResponse_httpStatus :: Lens' DeleteVirtualGatewayResponse Int

The response's http status code.

deleteVirtualGatewayResponse_virtualGateway :: Lens' DeleteVirtualGatewayResponse VirtualGatewayData

The virtual gateway that was deleted.