Amazonka.AppMesh.DeleteVirtualRouter

Description

Deletes an existing virtual router.

You must delete any routes associated with the virtual router before you can delete the router itself.

Creating a Request

data DeleteVirtualRouter

See: newDeleteVirtualRouter smart constructor.

Constructors

DeleteVirtualRouter'

Fields

meshOwner :: Maybe Text

The AWS IAM account ID of the service mesh owner. If the account ID is not your own, then it's the ID of the account that shared the mesh with your account. For more information about mesh sharing, see Working with shared meshes.

meshName :: Text

The name of the service mesh to delete the virtual router in.

virtualRouterName :: Text

The name of the virtual router to delete.

newDeleteVirtualRouter

Arguments

:: Text
$sel:meshName:DeleteVirtualRouter'

-> Text
$sel:virtualRouterName:DeleteVirtualRouter'

-> DeleteVirtualRouter

Create a value of DeleteVirtualRouter with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:meshOwner:DeleteVirtualRouter', deleteVirtualRouter_meshOwner - The AWS IAM account ID of the service mesh owner. If the account ID is not your own, then it's the ID of the account that shared the mesh with your account. For more information about mesh sharing, see Working with shared meshes.

$sel:meshName:DeleteVirtualRouter', deleteVirtualRouter_meshName - The name of the service mesh to delete the virtual router in.

$sel:virtualRouterName:DeleteVirtualRouter', deleteVirtualRouter_virtualRouterName - The name of the virtual router to delete.

Request Lenses

deleteVirtualRouter_meshOwner :: Lens' DeleteVirtualRouter (Maybe Text)

The AWS IAM account ID of the service mesh owner. If the account ID is not your own, then it's the ID of the account that shared the mesh with your account. For more information about mesh sharing, see Working with shared meshes.

deleteVirtualRouter_meshName :: Lens' DeleteVirtualRouter Text

The name of the service mesh to delete the virtual router in.

deleteVirtualRouter_virtualRouterName :: Lens' DeleteVirtualRouter Text

The name of the virtual router to delete.

Destructuring the Response

data DeleteVirtualRouterResponse

See: newDeleteVirtualRouterResponse smart constructor.

Constructors

DeleteVirtualRouterResponse'

Fields

httpStatus :: Int

The response's http status code.

virtualRouter :: VirtualRouterData

The virtual router that was deleted.

newDeleteVirtualRouterResponse

Arguments

:: Int
$sel:httpStatus:DeleteVirtualRouterResponse'

-> VirtualRouterData
$sel:virtualRouter:DeleteVirtualRouterResponse'

-> DeleteVirtualRouterResponse

Create a value of DeleteVirtualRouterResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:httpStatus:DeleteVirtualRouterResponse', deleteVirtualRouterResponse_httpStatus - The response's http status code.

$sel:virtualRouter:DeleteVirtualRouterResponse', deleteVirtualRouterResponse_virtualRouter - The virtual router that was deleted.

Response Lenses

deleteVirtualRouterResponse_httpStatus :: Lens' DeleteVirtualRouterResponse Int

The response's http status code.

deleteVirtualRouterResponse_virtualRouter :: Lens' DeleteVirtualRouterResponse VirtualRouterData

The virtual router that was deleted.