Amazonka.AppMesh.DescribeMesh

Description

Describes an existing service mesh.

Creating a Request

data DescribeMesh

See: newDescribeMesh smart constructor.

Constructors

DescribeMesh'

Fields

meshOwner :: Maybe Text

The AWS IAM account ID of the service mesh owner. If the account ID is not your own, then it's the ID of the account that shared the mesh with your account. For more information about mesh sharing, see Working with shared meshes.

meshName :: Text

The name of the service mesh to describe.

newDescribeMesh

Arguments

:: Text
$sel:meshName:DescribeMesh'

-> DescribeMesh

Create a value of DescribeMesh with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:meshOwner:DescribeMesh', describeMesh_meshOwner - The AWS IAM account ID of the service mesh owner. If the account ID is not your own, then it's the ID of the account that shared the mesh with your account. For more information about mesh sharing, see Working with shared meshes.

$sel:meshName:DescribeMesh', describeMesh_meshName - The name of the service mesh to describe.

Request Lenses

describeMesh_meshOwner :: Lens' DescribeMesh (Maybe Text)

The AWS IAM account ID of the service mesh owner. If the account ID is not your own, then it's the ID of the account that shared the mesh with your account. For more information about mesh sharing, see Working with shared meshes.

describeMesh_meshName :: Lens' DescribeMesh Text

The name of the service mesh to describe.

Destructuring the Response

data DescribeMeshResponse

See: newDescribeMeshResponse smart constructor.

Constructors

DescribeMeshResponse'

Fields

httpStatus :: Int

The response's http status code.

mesh :: MeshData

The full description of your service mesh.

newDescribeMeshResponse

Arguments

:: Int
$sel:httpStatus:DescribeMeshResponse'

-> MeshData
$sel:mesh:DescribeMeshResponse'

-> DescribeMeshResponse

Create a value of DescribeMeshResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:httpStatus:DescribeMeshResponse', describeMeshResponse_httpStatus - The response's http status code.

$sel:mesh:DescribeMeshResponse', describeMeshResponse_mesh - The full description of your service mesh.

Response Lenses

describeMeshResponse_httpStatus :: Lens' DescribeMeshResponse Int

The response's http status code.

describeMeshResponse_mesh :: Lens' DescribeMeshResponse MeshData

The full description of your service mesh.