Amazonka.AppMesh.DescribeVirtualNode

Description

Describes an existing virtual node.

Creating a Request

data DescribeVirtualNode

See: newDescribeVirtualNode smart constructor.

Constructors

DescribeVirtualNode'

Fields

meshOwner :: Maybe Text

The AWS IAM account ID of the service mesh owner. If the account ID is not your own, then it's the ID of the account that shared the mesh with your account. For more information about mesh sharing, see Working with shared meshes.

meshName :: Text

The name of the service mesh that the virtual node resides in.

virtualNodeName :: Text

The name of the virtual node to describe.

newDescribeVirtualNode

Arguments

:: Text
$sel:meshName:DescribeVirtualNode'

-> Text
$sel:virtualNodeName:DescribeVirtualNode'

-> DescribeVirtualNode

Create a value of DescribeVirtualNode with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:meshOwner:DescribeVirtualNode', describeVirtualNode_meshOwner - The AWS IAM account ID of the service mesh owner. If the account ID is not your own, then it's the ID of the account that shared the mesh with your account. For more information about mesh sharing, see Working with shared meshes.

$sel:meshName:DescribeVirtualNode', describeVirtualNode_meshName - The name of the service mesh that the virtual node resides in.

$sel:virtualNodeName:DescribeVirtualNode', describeVirtualNode_virtualNodeName - The name of the virtual node to describe.

Request Lenses

describeVirtualNode_meshOwner :: Lens' DescribeVirtualNode (Maybe Text)

The AWS IAM account ID of the service mesh owner. If the account ID is not your own, then it's the ID of the account that shared the mesh with your account. For more information about mesh sharing, see Working with shared meshes.

describeVirtualNode_meshName :: Lens' DescribeVirtualNode Text

The name of the service mesh that the virtual node resides in.

describeVirtualNode_virtualNodeName :: Lens' DescribeVirtualNode Text

The name of the virtual node to describe.

Destructuring the Response

data DescribeVirtualNodeResponse

See: newDescribeVirtualNodeResponse smart constructor.

Constructors

DescribeVirtualNodeResponse'

Fields

httpStatus :: Int

The response's http status code.

virtualNode :: VirtualNodeData

The full description of your virtual node.

newDescribeVirtualNodeResponse

Arguments

:: Int
$sel:httpStatus:DescribeVirtualNodeResponse'

-> VirtualNodeData
$sel:virtualNode:DescribeVirtualNodeResponse'

-> DescribeVirtualNodeResponse

Create a value of DescribeVirtualNodeResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:httpStatus:DescribeVirtualNodeResponse', describeVirtualNodeResponse_httpStatus - The response's http status code.

$sel:virtualNode:DescribeVirtualNodeResponse', describeVirtualNodeResponse_virtualNode - The full description of your virtual node.

Response Lenses

describeVirtualNodeResponse_httpStatus :: Lens' DescribeVirtualNodeResponse Int

The response's http status code.

describeVirtualNodeResponse_virtualNode :: Lens' DescribeVirtualNodeResponse VirtualNodeData

The full description of your virtual node.