Amazonka.AppMesh.DescribeVirtualRouter

Description

Describes an existing virtual router.

Creating a Request

data DescribeVirtualRouter

See: newDescribeVirtualRouter smart constructor.

Constructors

DescribeVirtualRouter'

Fields

meshOwner :: Maybe Text

The AWS IAM account ID of the service mesh owner. If the account ID is not your own, then it's the ID of the account that shared the mesh with your account. For more information about mesh sharing, see Working with shared meshes.

meshName :: Text

The name of the service mesh that the virtual router resides in.

virtualRouterName :: Text

The name of the virtual router to describe.

newDescribeVirtualRouter

Arguments

:: Text
$sel:meshName:DescribeVirtualRouter'

-> Text
$sel:virtualRouterName:DescribeVirtualRouter'

-> DescribeVirtualRouter

Create a value of DescribeVirtualRouter with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:meshOwner:DescribeVirtualRouter', describeVirtualRouter_meshOwner - The AWS IAM account ID of the service mesh owner. If the account ID is not your own, then it's the ID of the account that shared the mesh with your account. For more information about mesh sharing, see Working with shared meshes.

$sel:meshName:DescribeVirtualRouter', describeVirtualRouter_meshName - The name of the service mesh that the virtual router resides in.

$sel:virtualRouterName:DescribeVirtualRouter', describeVirtualRouter_virtualRouterName - The name of the virtual router to describe.

Request Lenses

describeVirtualRouter_meshOwner :: Lens' DescribeVirtualRouter (Maybe Text)

The AWS IAM account ID of the service mesh owner. If the account ID is not your own, then it's the ID of the account that shared the mesh with your account. For more information about mesh sharing, see Working with shared meshes.

describeVirtualRouter_meshName :: Lens' DescribeVirtualRouter Text

The name of the service mesh that the virtual router resides in.

describeVirtualRouter_virtualRouterName :: Lens' DescribeVirtualRouter Text

The name of the virtual router to describe.

Destructuring the Response

data DescribeVirtualRouterResponse

See: newDescribeVirtualRouterResponse smart constructor.

Constructors

DescribeVirtualRouterResponse'

Fields

httpStatus :: Int

The response's http status code.

virtualRouter :: VirtualRouterData

The full description of your virtual router.

newDescribeVirtualRouterResponse

Arguments

:: Int
$sel:httpStatus:DescribeVirtualRouterResponse'

-> VirtualRouterData
$sel:virtualRouter:DescribeVirtualRouterResponse'

-> DescribeVirtualRouterResponse

Create a value of DescribeVirtualRouterResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:httpStatus:DescribeVirtualRouterResponse', describeVirtualRouterResponse_httpStatus - The response's http status code.

$sel:virtualRouter:DescribeVirtualRouterResponse', describeVirtualRouterResponse_virtualRouter - The full description of your virtual router.

Response Lenses

describeVirtualRouterResponse_httpStatus :: Lens' DescribeVirtualRouterResponse Int

The response's http status code.

describeVirtualRouterResponse_virtualRouter :: Lens' DescribeVirtualRouterResponse VirtualRouterData

The full description of your virtual router.