Amazonka.AppMesh.ListRoutes

Description

Returns a list of existing routes in a service mesh.

This operation returns paginated results.

Creating a Request

data ListRoutes

See: newListRoutes smart constructor.

Constructors

ListRoutes'

Fields

meshOwner :: Maybe Text

The AWS IAM account ID of the service mesh owner. If the account ID is not your own, then it's the ID of the account that shared the mesh with your account. For more information about mesh sharing, see Working with shared meshes.

nextToken :: Maybe Text

The nextToken value returned from a previous paginated ListRoutes request where limit was used and the results exceeded the value of that parameter. Pagination continues from the end of the previous results that returned the nextToken value.

limit :: Maybe Natural

The maximum number of results returned by ListRoutes in paginated output. When you use this parameter, ListRoutes returns only limit results in a single page along with a nextToken response element. You can see the remaining results of the initial request by sending another ListRoutes request with the returned nextToken value. This value can be between 1 and 100. If you don't use this parameter, ListRoutes returns up to 100 results and a nextToken value if applicable.

meshName :: Text

The name of the service mesh to list routes in.

virtualRouterName :: Text

The name of the virtual router to list routes in.

newListRoutes

Arguments

:: Text
$sel:meshName:ListRoutes'

-> Text
$sel:virtualRouterName:ListRoutes'

-> ListRoutes

Create a value of ListRoutes with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:meshOwner:ListRoutes', listRoutes_meshOwner - The AWS IAM account ID of the service mesh owner. If the account ID is not your own, then it's the ID of the account that shared the mesh with your account. For more information about mesh sharing, see Working with shared meshes.

$sel:nextToken:ListRoutes', listRoutes_nextToken - The nextToken value returned from a previous paginated ListRoutes request where limit was used and the results exceeded the value of that parameter. Pagination continues from the end of the previous results that returned the nextToken value.

$sel:limit:ListRoutes', listRoutes_limit - The maximum number of results returned by ListRoutes in paginated output. When you use this parameter, ListRoutes returns only limit results in a single page along with a nextToken response element. You can see the remaining results of the initial request by sending another ListRoutes request with the returned nextToken value. This value can be between 1 and 100. If you don't use this parameter, ListRoutes returns up to 100 results and a nextToken value if applicable.

$sel:meshName:ListRoutes', listRoutes_meshName - The name of the service mesh to list routes in.

$sel:virtualRouterName:ListRoutes', listRoutes_virtualRouterName - The name of the virtual router to list routes in.

Request Lenses

listRoutes_meshOwner :: Lens' ListRoutes (Maybe Text)

The AWS IAM account ID of the service mesh owner. If the account ID is not your own, then it's the ID of the account that shared the mesh with your account. For more information about mesh sharing, see Working with shared meshes.

listRoutes_nextToken :: Lens' ListRoutes (Maybe Text)

The nextToken value returned from a previous paginated ListRoutes request where limit was used and the results exceeded the value of that parameter. Pagination continues from the end of the previous results that returned the nextToken value.

listRoutes_limit :: Lens' ListRoutes (Maybe Natural)

The maximum number of results returned by ListRoutes in paginated output. When you use this parameter, ListRoutes returns only limit results in a single page along with a nextToken response element. You can see the remaining results of the initial request by sending another ListRoutes request with the returned nextToken value. This value can be between 1 and 100. If you don't use this parameter, ListRoutes returns up to 100 results and a nextToken value if applicable.

listRoutes_meshName :: Lens' ListRoutes Text

The name of the service mesh to list routes in.

listRoutes_virtualRouterName :: Lens' ListRoutes Text

The name of the virtual router to list routes in.

Destructuring the Response

data ListRoutesResponse

See: newListRoutesResponse smart constructor.

Constructors

ListRoutesResponse'

Fields

nextToken :: Maybe Text

The nextToken value to include in a future ListRoutes request. When the results of a ListRoutes request exceed limit, you can use this value to retrieve the next page of results. This value is null when there are no more results to return.

httpStatus :: Int

The response's http status code.

routes :: [RouteRef]

The list of existing routes for the specified service mesh and virtual router.

newListRoutesResponse

Arguments

:: Int
$sel:httpStatus:ListRoutesResponse'

-> ListRoutesResponse

Create a value of ListRoutesResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:nextToken:ListRoutes', listRoutesResponse_nextToken - The nextToken value to include in a future ListRoutes request. When the results of a ListRoutes request exceed limit, you can use this value to retrieve the next page of results. This value is null when there are no more results to return.

$sel:httpStatus:ListRoutesResponse', listRoutesResponse_httpStatus - The response's http status code.

$sel:routes:ListRoutesResponse', listRoutesResponse_routes - The list of existing routes for the specified service mesh and virtual router.

Response Lenses

listRoutesResponse_nextToken :: Lens' ListRoutesResponse (Maybe Text)

The nextToken value to include in a future ListRoutes request. When the results of a ListRoutes request exceed limit, you can use this value to retrieve the next page of results. This value is null when there are no more results to return.

listRoutesResponse_httpStatus :: Lens' ListRoutesResponse Int

The response's http status code.

listRoutesResponse_routes :: Lens' ListRoutesResponse [RouteRef]

The list of existing routes for the specified service mesh and virtual router.