Amazonka.AppMesh.Types.AwsCloudMapServiceDiscovery

Description

Documentation

data AwsCloudMapServiceDiscovery

An object that represents the Cloud Map service discovery information for your virtual node.

Cloud Map is not available in the eu-south-1 Region.

See: newAwsCloudMapServiceDiscovery smart constructor.

Constructors

AwsCloudMapServiceDiscovery'

Fields

attributes :: Maybe [AwsCloudMapInstanceAttribute]

A string map that contains attributes with values that you can use to filter instances by any custom attribute that you specified when you registered the instance. Only instances that match all of the specified key/value pairs will be returned.

namespaceName :: Text

The name of the Cloud Map namespace to use.

serviceName :: Text

The name of the Cloud Map service to use.

newAwsCloudMapServiceDiscovery

Arguments

:: Text
$sel:namespaceName:AwsCloudMapServiceDiscovery'

-> Text
$sel:serviceName:AwsCloudMapServiceDiscovery'

-> AwsCloudMapServiceDiscovery

Create a value of AwsCloudMapServiceDiscovery with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:attributes:AwsCloudMapServiceDiscovery', awsCloudMapServiceDiscovery_attributes - A string map that contains attributes with values that you can use to filter instances by any custom attribute that you specified when you registered the instance. Only instances that match all of the specified key/value pairs will be returned.

$sel:namespaceName:AwsCloudMapServiceDiscovery', awsCloudMapServiceDiscovery_namespaceName - The name of the Cloud Map namespace to use.

$sel:serviceName:AwsCloudMapServiceDiscovery', awsCloudMapServiceDiscovery_serviceName - The name of the Cloud Map service to use.

awsCloudMapServiceDiscovery_attributes :: Lens' AwsCloudMapServiceDiscovery (Maybe [AwsCloudMapInstanceAttribute])

A string map that contains attributes with values that you can use to filter instances by any custom attribute that you specified when you registered the instance. Only instances that match all of the specified key/value pairs will be returned.

awsCloudMapServiceDiscovery_namespaceName :: Lens' AwsCloudMapServiceDiscovery Text

The name of the Cloud Map namespace to use.

awsCloudMapServiceDiscovery_serviceName :: Lens' AwsCloudMapServiceDiscovery Text

The name of the Cloud Map service to use.