Amazonka.AppMesh.Types.ClientPolicyTls

Description

Documentation

data ClientPolicyTls

A reference to an object that represents a Transport Layer Security (TLS) client policy.

See: newClientPolicyTls smart constructor.

Constructors

ClientPolicyTls'

Fields

ports :: Maybe [Natural]

One or more ports that the policy is enforced for.

certificate :: Maybe ClientTlsCertificate

A reference to an object that represents a client's TLS certificate.

enforce :: Maybe Bool

Whether the policy is enforced. The default is True, if a value isn't specified.

validation :: TlsValidationContext

A reference to an object that represents a TLS validation context.

newClientPolicyTls

Arguments

:: TlsValidationContext
$sel:validation:ClientPolicyTls'

-> ClientPolicyTls

Create a value of ClientPolicyTls with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:ports:ClientPolicyTls', clientPolicyTls_ports - One or more ports that the policy is enforced for.

$sel:certificate:ClientPolicyTls', clientPolicyTls_certificate - A reference to an object that represents a client's TLS certificate.

$sel:enforce:ClientPolicyTls', clientPolicyTls_enforce - Whether the policy is enforced. The default is True, if a value isn't specified.

$sel:validation:ClientPolicyTls', clientPolicyTls_validation - A reference to an object that represents a TLS validation context.

clientPolicyTls_ports :: Lens' ClientPolicyTls (Maybe [Natural])

One or more ports that the policy is enforced for.

clientPolicyTls_certificate :: Lens' ClientPolicyTls (Maybe ClientTlsCertificate)

A reference to an object that represents a client's TLS certificate.

clientPolicyTls_enforce :: Lens' ClientPolicyTls (Maybe Bool)

Whether the policy is enforced. The default is True, if a value isn't specified.

clientPolicyTls_validation :: Lens' ClientPolicyTls TlsValidationContext

A reference to an object that represents a TLS validation context.