Amazonka.AppMesh.Types.GrpcGatewayRouteAction

Description

Documentation

data GrpcGatewayRouteAction

An object that represents the action to take if a match is determined.

See: newGrpcGatewayRouteAction smart constructor.

Constructors

GrpcGatewayRouteAction'

Fields

rewrite :: Maybe GrpcGatewayRouteRewrite

The gateway route action to rewrite.

target :: GatewayRouteTarget

An object that represents the target that traffic is routed to when a request matches the gateway route.

newGrpcGatewayRouteAction

Arguments

:: GatewayRouteTarget
$sel:target:GrpcGatewayRouteAction'

-> GrpcGatewayRouteAction

Create a value of GrpcGatewayRouteAction with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:rewrite:GrpcGatewayRouteAction', grpcGatewayRouteAction_rewrite - The gateway route action to rewrite.

$sel:target:GrpcGatewayRouteAction', grpcGatewayRouteAction_target - An object that represents the target that traffic is routed to when a request matches the gateway route.

grpcGatewayRouteAction_rewrite :: Lens' GrpcGatewayRouteAction (Maybe GrpcGatewayRouteRewrite)

The gateway route action to rewrite.

grpcGatewayRouteAction_target :: Lens' GrpcGatewayRouteAction GatewayRouteTarget

An object that represents the target that traffic is routed to when a request matches the gateway route.