Amazonka.AppMesh.Types.GrpcMetadataMatchMethod

Description

Documentation

data GrpcMetadataMatchMethod

An object representing the method header to be matched.

See: newGrpcMetadataMatchMethod smart constructor.

Constructors

GrpcMetadataMatchMethod'

Fields

suffix :: Maybe Text

The specified ending characters of the method header to match on.

regex :: Maybe Text

The regex used to match the method header.

prefix :: Maybe Text

The specified beginning characters of the method header to be matched on.

range :: Maybe MatchRange

exact :: Maybe Text

The exact method header to be matched on.

newGrpcMetadataMatchMethod :: GrpcMetadataMatchMethod

Create a value of GrpcMetadataMatchMethod with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:suffix:GrpcMetadataMatchMethod', grpcMetadataMatchMethod_suffix - The specified ending characters of the method header to match on.

$sel:regex:GrpcMetadataMatchMethod', grpcMetadataMatchMethod_regex - The regex used to match the method header.

$sel:prefix:GrpcMetadataMatchMethod', grpcMetadataMatchMethod_prefix - The specified beginning characters of the method header to be matched on.

$sel:range:GrpcMetadataMatchMethod', grpcMetadataMatchMethod_range - Undocumented member.

$sel:exact:GrpcMetadataMatchMethod', grpcMetadataMatchMethod_exact - The exact method header to be matched on.

grpcMetadataMatchMethod_suffix :: Lens' GrpcMetadataMatchMethod (Maybe Text)

The specified ending characters of the method header to match on.

grpcMetadataMatchMethod_regex :: Lens' GrpcMetadataMatchMethod (Maybe Text)

The regex used to match the method header.

grpcMetadataMatchMethod_prefix :: Lens' GrpcMetadataMatchMethod (Maybe Text)

The specified beginning characters of the method header to be matched on.

grpcMetadataMatchMethod_range :: Lens' GrpcMetadataMatchMethod (Maybe MatchRange)

Undocumented member.

grpcMetadataMatchMethod_exact :: Lens' GrpcMetadataMatchMethod (Maybe Text)

The exact method header to be matched on.