Amazonka.AppMesh.Types.HeaderMatchMethod

Description

Documentation

data HeaderMatchMethod

An object that represents the method and value to match with the header value sent in a request. Specify one match method.

See: newHeaderMatchMethod smart constructor.

Constructors

HeaderMatchMethod'

Fields

suffix :: Maybe Text

The value sent by the client must end with the specified characters.

regex :: Maybe Text

The value sent by the client must include the specified characters.

prefix :: Maybe Text

The value sent by the client must begin with the specified characters.

range :: Maybe MatchRange

An object that represents the range of values to match on.

exact :: Maybe Text

The value sent by the client must match the specified value exactly.

newHeaderMatchMethod :: HeaderMatchMethod

Create a value of HeaderMatchMethod with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:suffix:HeaderMatchMethod', headerMatchMethod_suffix - The value sent by the client must end with the specified characters.

$sel:regex:HeaderMatchMethod', headerMatchMethod_regex - The value sent by the client must include the specified characters.

$sel:prefix:HeaderMatchMethod', headerMatchMethod_prefix - The value sent by the client must begin with the specified characters.

$sel:range:HeaderMatchMethod', headerMatchMethod_range - An object that represents the range of values to match on.

$sel:exact:HeaderMatchMethod', headerMatchMethod_exact - The value sent by the client must match the specified value exactly.

headerMatchMethod_suffix :: Lens' HeaderMatchMethod (Maybe Text)

The value sent by the client must end with the specified characters.

headerMatchMethod_regex :: Lens' HeaderMatchMethod (Maybe Text)

The value sent by the client must include the specified characters.

headerMatchMethod_prefix :: Lens' HeaderMatchMethod (Maybe Text)

The value sent by the client must begin with the specified characters.

headerMatchMethod_range :: Lens' HeaderMatchMethod (Maybe MatchRange)

An object that represents the range of values to match on.

headerMatchMethod_exact :: Lens' HeaderMatchMethod (Maybe Text)

The value sent by the client must match the specified value exactly.