Amazonka.AppMesh.Types.HttpGatewayRouteMatch

Description

Documentation

data HttpGatewayRouteMatch

An object that represents the criteria for determining a request match.

See: newHttpGatewayRouteMatch smart constructor.

Constructors

HttpGatewayRouteMatch'

Fields

hostname :: Maybe GatewayRouteHostnameMatch

The host name to match on.

path :: Maybe HttpPathMatch

The path to match on.

prefix :: Maybe Text

Specifies the path to match requests with. This parameter must always start with /, which by itself matches all requests to the virtual service name. You can also match for path-based routing of requests. For example, if your virtual service name is my-service.local and you want the route to match requests to my-service.local/metrics, your prefix should be /metrics.

queryParameters :: Maybe (NonEmpty HttpQueryParameter)

The query parameter to match on.

headers :: Maybe (NonEmpty HttpGatewayRouteHeader)

The client request headers to match on.

method :: Maybe HttpMethod

The method to match on.

newHttpGatewayRouteMatch :: HttpGatewayRouteMatch

Create a value of HttpGatewayRouteMatch with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:hostname:HttpGatewayRouteMatch', httpGatewayRouteMatch_hostname - The host name to match on.

$sel:path:HttpGatewayRouteMatch', httpGatewayRouteMatch_path - The path to match on.

$sel:prefix:HttpGatewayRouteMatch', httpGatewayRouteMatch_prefix - Specifies the path to match requests with. This parameter must always start with /, which by itself matches all requests to the virtual service name. You can also match for path-based routing of requests. For example, if your virtual service name is my-service.local and you want the route to match requests to my-service.local/metrics, your prefix should be /metrics.

$sel:queryParameters:HttpGatewayRouteMatch', httpGatewayRouteMatch_queryParameters - The query parameter to match on.

$sel:headers:HttpGatewayRouteMatch', httpGatewayRouteMatch_headers - The client request headers to match on.

$sel:method:HttpGatewayRouteMatch', httpGatewayRouteMatch_method - The method to match on.

httpGatewayRouteMatch_hostname :: Lens' HttpGatewayRouteMatch (Maybe GatewayRouteHostnameMatch)

The host name to match on.

httpGatewayRouteMatch_path :: Lens' HttpGatewayRouteMatch (Maybe HttpPathMatch)

The path to match on.

httpGatewayRouteMatch_prefix :: Lens' HttpGatewayRouteMatch (Maybe Text)

Specifies the path to match requests with. This parameter must always start with /, which by itself matches all requests to the virtual service name. You can also match for path-based routing of requests. For example, if your virtual service name is my-service.local and you want the route to match requests to my-service.local/metrics, your prefix should be /metrics.

httpGatewayRouteMatch_queryParameters :: Lens' HttpGatewayRouteMatch (Maybe (NonEmpty HttpQueryParameter))

The query parameter to match on.

httpGatewayRouteMatch_headers :: Lens' HttpGatewayRouteMatch (Maybe (NonEmpty HttpGatewayRouteHeader))

The client request headers to match on.

httpGatewayRouteMatch_method :: Lens' HttpGatewayRouteMatch (Maybe HttpMethod)

The method to match on.