Amazonka.AppMesh.Types.Listener

Description

Documentation

data Listener

An object that represents a listener for a virtual node.

See: newListener smart constructor.

Constructors

Listener'

Fields

healthCheck :: Maybe HealthCheckPolicy

The health check information for the listener.

connectionPool :: Maybe VirtualNodeConnectionPool

The connection pool information for the listener.

tls :: Maybe ListenerTls

A reference to an object that represents the Transport Layer Security (TLS) properties for a listener.

outlierDetection :: Maybe OutlierDetection

The outlier detection information for the listener.

timeout :: Maybe ListenerTimeout

An object that represents timeouts for different protocols.

portMapping :: PortMapping

The port mapping information for the listener.

newListener

Arguments

:: PortMapping
$sel:portMapping:Listener'

-> Listener

Create a value of Listener with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:healthCheck:Listener', listener_healthCheck - The health check information for the listener.

$sel:connectionPool:Listener', listener_connectionPool - The connection pool information for the listener.

$sel:tls:Listener', listener_tls - A reference to an object that represents the Transport Layer Security (TLS) properties for a listener.

$sel:outlierDetection:Listener', listener_outlierDetection - The outlier detection information for the listener.

$sel:timeout:Listener', listener_timeout - An object that represents timeouts for different protocols.

$sel:portMapping:Listener', listener_portMapping - The port mapping information for the listener.

listener_healthCheck :: Lens' Listener (Maybe HealthCheckPolicy)

The health check information for the listener.

listener_connectionPool :: Lens' Listener (Maybe VirtualNodeConnectionPool)

The connection pool information for the listener.

listener_tls :: Lens' Listener (Maybe ListenerTls)

A reference to an object that represents the Transport Layer Security (TLS) properties for a listener.

listener_outlierDetection :: Lens' Listener (Maybe OutlierDetection)

The outlier detection information for the listener.

listener_timeout :: Lens' Listener (Maybe ListenerTimeout)

An object that represents timeouts for different protocols.

listener_portMapping :: Lens' Listener PortMapping

The port mapping information for the listener.