Amazonka.AppMesh.Types.ListenerTls

Description

Documentation

data ListenerTls

An object that represents the Transport Layer Security (TLS) properties for a listener.

See: newListenerTls smart constructor.

Constructors

ListenerTls'

Fields

validation :: Maybe ListenerTlsValidationContext

A reference to an object that represents a listener's Transport Layer Security (TLS) validation context.

certificate :: ListenerTlsCertificate

A reference to an object that represents a listener's Transport Layer Security (TLS) certificate.

mode :: ListenerTlsMode

Specify one of the following modes.

____STRICT – Listener only accepts connections with TLS enabled.

____PERMISSIVE – Listener accepts connections with or without TLS enabled.

____DISABLED – Listener only accepts connections without TLS.

newListenerTls

Arguments

:: ListenerTlsCertificate
$sel:certificate:ListenerTls'

-> ListenerTlsMode
$sel:mode:ListenerTls'

-> ListenerTls

Create a value of ListenerTls with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:validation:ListenerTls', listenerTls_validation - A reference to an object that represents a listener's Transport Layer Security (TLS) validation context.

$sel:certificate:ListenerTls', listenerTls_certificate - A reference to an object that represents a listener's Transport Layer Security (TLS) certificate.

$sel:mode:ListenerTls', listenerTls_mode - Specify one of the following modes.

____STRICT – Listener only accepts connections with TLS enabled.

____PERMISSIVE – Listener accepts connections with or without TLS enabled.

____DISABLED – Listener only accepts connections without TLS.

listenerTls_validation :: Lens' ListenerTls (Maybe ListenerTlsValidationContext)

A reference to an object that represents a listener's Transport Layer Security (TLS) validation context.

listenerTls_certificate :: Lens' ListenerTls ListenerTlsCertificate

A reference to an object that represents a listener's Transport Layer Security (TLS) certificate.

listenerTls_mode :: Lens' ListenerTls ListenerTlsMode

Specify one of the following modes.

____STRICT – Listener only accepts connections with TLS enabled.

____PERMISSIVE – Listener accepts connections with or without TLS enabled.

____DISABLED – Listener only accepts connections without TLS.