Amazonka.AppMesh.Types.ListenerTlsCertificate

Description

Documentation

data ListenerTlsCertificate

An object that represents a listener's Transport Layer Security (TLS) certificate.

See: newListenerTlsCertificate smart constructor.

Constructors

ListenerTlsCertificate'

Fields

acm :: Maybe ListenerTlsAcmCertificate

A reference to an object that represents an AWS Certicate Manager (ACM) certificate.

sds :: Maybe ListenerTlsSdsCertificate

A reference to an object that represents a listener's Secret Discovery Service certificate.

file :: Maybe ListenerTlsFileCertificate

A reference to an object that represents a local file certificate.

newListenerTlsCertificate :: ListenerTlsCertificate

Create a value of ListenerTlsCertificate with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:acm:ListenerTlsCertificate', listenerTlsCertificate_acm - A reference to an object that represents an AWS Certicate Manager (ACM) certificate.

$sel:sds:ListenerTlsCertificate', listenerTlsCertificate_sds - A reference to an object that represents a listener's Secret Discovery Service certificate.

$sel:file:ListenerTlsCertificate', listenerTlsCertificate_file - A reference to an object that represents a local file certificate.

listenerTlsCertificate_acm :: Lens' ListenerTlsCertificate (Maybe ListenerTlsAcmCertificate)

A reference to an object that represents an AWS Certicate Manager (ACM) certificate.

listenerTlsCertificate_sds :: Lens' ListenerTlsCertificate (Maybe ListenerTlsSdsCertificate)

A reference to an object that represents a listener's Secret Discovery Service certificate.

listenerTlsCertificate_file :: Lens' ListenerTlsCertificate (Maybe ListenerTlsFileCertificate)

A reference to an object that represents a local file certificate.