Amazonka.AppMesh.Types.MeshRef

Description

Documentation

data MeshRef

An object that represents a service mesh returned by a list operation.

See: newMeshRef smart constructor.

Constructors

MeshRef'

Fields

arn :: Text

The full Amazon Resource Name (ARN) of the service mesh.

createdAt :: POSIX

The Unix epoch timestamp in seconds for when the resource was created.

lastUpdatedAt :: POSIX

The Unix epoch timestamp in seconds for when the resource was last updated.

meshName :: Text

The name of the service mesh.

meshOwner :: Text

The AWS IAM account ID of the service mesh owner. If the account ID is not your own, then it's the ID of the account that shared the mesh with your account. For more information about mesh sharing, see Working with shared meshes.

resourceOwner :: Text

The AWS IAM account ID of the resource owner. If the account ID is not your own, then it's the ID of the mesh owner or of another account that the mesh is shared with. For more information about mesh sharing, see Working with shared meshes.

version :: Integer

The version of the resource. Resources are created at version 1, and this version is incremented each time that they're updated.

newMeshRef

Arguments

:: Text
$sel:arn:MeshRef'

-> UTCTime
$sel:createdAt:MeshRef'

-> UTCTime
$sel:lastUpdatedAt:MeshRef'

-> Text
$sel:meshName:MeshRef'

-> Text
$sel:meshOwner:MeshRef'

-> Text
$sel:resourceOwner:MeshRef'

-> Integer
$sel:version:MeshRef'

-> MeshRef

Create a value of MeshRef with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:arn:MeshRef', meshRef_arn - The full Amazon Resource Name (ARN) of the service mesh.

$sel:createdAt:MeshRef', meshRef_createdAt - The Unix epoch timestamp in seconds for when the resource was created.

$sel:lastUpdatedAt:MeshRef', meshRef_lastUpdatedAt - The Unix epoch timestamp in seconds for when the resource was last updated.

$sel:meshName:MeshRef', meshRef_meshName - The name of the service mesh.

$sel:meshOwner:MeshRef', meshRef_meshOwner - The AWS IAM account ID of the service mesh owner. If the account ID is not your own, then it's the ID of the account that shared the mesh with your account. For more information about mesh sharing, see Working with shared meshes.

$sel:resourceOwner:MeshRef', meshRef_resourceOwner - The AWS IAM account ID of the resource owner. If the account ID is not your own, then it's the ID of the mesh owner or of another account that the mesh is shared with. For more information about mesh sharing, see Working with shared meshes.

$sel:version:MeshRef', meshRef_version - The version of the resource. Resources are created at version 1, and this version is incremented each time that they're updated.

meshRef_arn :: Lens' MeshRef Text

The full Amazon Resource Name (ARN) of the service mesh.

meshRef_createdAt :: Lens' MeshRef UTCTime

The Unix epoch timestamp in seconds for when the resource was created.

meshRef_lastUpdatedAt :: Lens' MeshRef UTCTime

The Unix epoch timestamp in seconds for when the resource was last updated.

meshRef_meshName :: Lens' MeshRef Text

The name of the service mesh.

meshRef_meshOwner :: Lens' MeshRef Text

The AWS IAM account ID of the service mesh owner. If the account ID is not your own, then it's the ID of the account that shared the mesh with your account. For more information about mesh sharing, see Working with shared meshes.

meshRef_resourceOwner :: Lens' MeshRef Text

The AWS IAM account ID of the resource owner. If the account ID is not your own, then it's the ID of the mesh owner or of another account that the mesh is shared with. For more information about mesh sharing, see Working with shared meshes.

meshRef_version :: Lens' MeshRef Integer

The version of the resource. Resources are created at version 1, and this version is incremented each time that they're updated.