Amazonka.AppMesh.Types.OutlierDetection

Description

Documentation

data OutlierDetection

An object that represents the outlier detection for a virtual node's listener.

See: newOutlierDetection smart constructor.

Constructors

OutlierDetection'

Fields

baseEjectionDuration :: Duration

The base amount of time for which a host is ejected.

interval :: Duration

The time interval between ejection sweep analysis.

maxEjectionPercent :: Natural

Maximum percentage of hosts in load balancing pool for upstream service that can be ejected. Will eject at least one host regardless of the value.

maxServerErrors :: Natural

Number of consecutive 5xx errors required for ejection.

newOutlierDetection

Arguments

:: Duration
$sel:baseEjectionDuration:OutlierDetection'

-> Duration
$sel:interval:OutlierDetection'

-> Natural
$sel:maxEjectionPercent:OutlierDetection'

-> Natural
$sel:maxServerErrors:OutlierDetection'

-> OutlierDetection

Create a value of OutlierDetection with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:baseEjectionDuration:OutlierDetection', outlierDetection_baseEjectionDuration - The base amount of time for which a host is ejected.

$sel:interval:OutlierDetection', outlierDetection_interval - The time interval between ejection sweep analysis.

$sel:maxEjectionPercent:OutlierDetection', outlierDetection_maxEjectionPercent - Maximum percentage of hosts in load balancing pool for upstream service that can be ejected. Will eject at least one host regardless of the value.

$sel:maxServerErrors:OutlierDetection', outlierDetection_maxServerErrors - Number of consecutive 5xx errors required for ejection.

outlierDetection_baseEjectionDuration :: Lens' OutlierDetection Duration

The base amount of time for which a host is ejected.

outlierDetection_interval :: Lens' OutlierDetection Duration

The time interval between ejection sweep analysis.

outlierDetection_maxEjectionPercent :: Lens' OutlierDetection Natural

Maximum percentage of hosts in load balancing pool for upstream service that can be ejected. Will eject at least one host regardless of the value.

outlierDetection_maxServerErrors :: Lens' OutlierDetection Natural

Number of consecutive 5xx errors required for ejection.