Amazonka.AppMesh.Types.ResourceMetadata

Description

Documentation

data ResourceMetadata

An object that represents metadata for a resource.

See: newResourceMetadata smart constructor.

Constructors

ResourceMetadata'

Fields

arn :: Text

The full Amazon Resource Name (ARN) for the resource.

createdAt :: POSIX

The Unix epoch timestamp in seconds for when the resource was created.

lastUpdatedAt :: POSIX

The Unix epoch timestamp in seconds for when the resource was last updated.

meshOwner :: Text

The AWS IAM account ID of the service mesh owner. If the account ID is not your own, then it's the ID of the account that shared the mesh with your account. For more information about mesh sharing, see Working with shared meshes.

resourceOwner :: Text

The AWS IAM account ID of the resource owner. If the account ID is not your own, then it's the ID of the mesh owner or of another account that the mesh is shared with. For more information about mesh sharing, see Working with shared meshes.

uid :: Text

The unique identifier for the resource.

version :: Integer

The version of the resource. Resources are created at version 1, and this version is incremented each time that they're updated.

newResourceMetadata

Arguments

:: Text
$sel:arn:ResourceMetadata'

-> UTCTime
$sel:createdAt:ResourceMetadata'

-> UTCTime
$sel:lastUpdatedAt:ResourceMetadata'

-> Text
$sel:meshOwner:ResourceMetadata'

-> Text
$sel:resourceOwner:ResourceMetadata'

-> Text
$sel:uid:ResourceMetadata'

-> Integer
$sel:version:ResourceMetadata'

-> ResourceMetadata

Create a value of ResourceMetadata with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:arn:ResourceMetadata', resourceMetadata_arn - The full Amazon Resource Name (ARN) for the resource.

$sel:createdAt:ResourceMetadata', resourceMetadata_createdAt - The Unix epoch timestamp in seconds for when the resource was created.

$sel:lastUpdatedAt:ResourceMetadata', resourceMetadata_lastUpdatedAt - The Unix epoch timestamp in seconds for when the resource was last updated.

$sel:meshOwner:ResourceMetadata', resourceMetadata_meshOwner - The AWS IAM account ID of the service mesh owner. If the account ID is not your own, then it's the ID of the account that shared the mesh with your account. For more information about mesh sharing, see Working with shared meshes.

$sel:resourceOwner:ResourceMetadata', resourceMetadata_resourceOwner - The AWS IAM account ID of the resource owner. If the account ID is not your own, then it's the ID of the mesh owner or of another account that the mesh is shared with. For more information about mesh sharing, see Working with shared meshes.

$sel:uid:ResourceMetadata', resourceMetadata_uid - The unique identifier for the resource.

$sel:version:ResourceMetadata', resourceMetadata_version - The version of the resource. Resources are created at version 1, and this version is incremented each time that they're updated.

resourceMetadata_arn :: Lens' ResourceMetadata Text

The full Amazon Resource Name (ARN) for the resource.

resourceMetadata_createdAt :: Lens' ResourceMetadata UTCTime

The Unix epoch timestamp in seconds for when the resource was created.

resourceMetadata_lastUpdatedAt :: Lens' ResourceMetadata UTCTime

The Unix epoch timestamp in seconds for when the resource was last updated.

resourceMetadata_meshOwner :: Lens' ResourceMetadata Text

The AWS IAM account ID of the service mesh owner. If the account ID is not your own, then it's the ID of the account that shared the mesh with your account. For more information about mesh sharing, see Working with shared meshes.

resourceMetadata_resourceOwner :: Lens' ResourceMetadata Text

The AWS IAM account ID of the resource owner. If the account ID is not your own, then it's the ID of the mesh owner or of another account that the mesh is shared with. For more information about mesh sharing, see Working with shared meshes.

resourceMetadata_uid :: Lens' ResourceMetadata Text

The unique identifier for the resource.

resourceMetadata_version :: Lens' ResourceMetadata Integer

The version of the resource. Resources are created at version 1, and this version is incremented each time that they're updated.