Amazonka.AppMesh.Types.VirtualGatewayAccessLog

Description

Documentation

data VirtualGatewayAccessLog

The access log configuration for a virtual gateway.

See: newVirtualGatewayAccessLog smart constructor.

Constructors

VirtualGatewayAccessLog'

Fields

file :: Maybe VirtualGatewayFileAccessLog

The file object to send virtual gateway access logs to.

newVirtualGatewayAccessLog :: VirtualGatewayAccessLog

Create a value of VirtualGatewayAccessLog with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:file:VirtualGatewayAccessLog', virtualGatewayAccessLog_file - The file object to send virtual gateway access logs to.

virtualGatewayAccessLog_file :: Lens' VirtualGatewayAccessLog (Maybe VirtualGatewayFileAccessLog)

The file object to send virtual gateway access logs to.