Amazonka.AppMesh.Types.VirtualGatewayHealthCheckPolicy

Description

Documentation

data VirtualGatewayHealthCheckPolicy

An object that represents the health check policy for a virtual gateway's listener.

See: newVirtualGatewayHealthCheckPolicy smart constructor.

Constructors

VirtualGatewayHealthCheckPolicy'

Fields

path :: Maybe Text

The destination path for the health check request. This value is only used if the specified protocol is HTTP or HTTP/2. For any other protocol, this value is ignored.

port :: Maybe Natural

The destination port for the health check request. This port must match the port defined in the PortMapping for the listener.

healthyThreshold :: Natural

The number of consecutive successful health checks that must occur before declaring the listener healthy.

intervalMillis :: Natural

The time period in milliseconds between each health check execution.

protocol :: VirtualGatewayPortProtocol

The protocol for the health check request. If you specify grpc, then your service must conform to the GRPC Health Checking Protocol.

timeoutMillis :: Natural

The amount of time to wait when receiving a response from the health check, in milliseconds.

unhealthyThreshold :: Natural

The number of consecutive failed health checks that must occur before declaring a virtual gateway unhealthy.

newVirtualGatewayHealthCheckPolicy

Arguments

:: Natural
$sel:healthyThreshold:VirtualGatewayHealthCheckPolicy'

-> Natural
$sel:intervalMillis:VirtualGatewayHealthCheckPolicy'

-> VirtualGatewayPortProtocol
$sel:protocol:VirtualGatewayHealthCheckPolicy'

-> Natural
$sel:timeoutMillis:VirtualGatewayHealthCheckPolicy'

-> Natural
$sel:unhealthyThreshold:VirtualGatewayHealthCheckPolicy'

-> VirtualGatewayHealthCheckPolicy

Create a value of VirtualGatewayHealthCheckPolicy with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:path:VirtualGatewayHealthCheckPolicy', virtualGatewayHealthCheckPolicy_path - The destination path for the health check request. This value is only used if the specified protocol is HTTP or HTTP/2. For any other protocol, this value is ignored.

$sel:port:VirtualGatewayHealthCheckPolicy', virtualGatewayHealthCheckPolicy_port - The destination port for the health check request. This port must match the port defined in the PortMapping for the listener.

$sel:healthyThreshold:VirtualGatewayHealthCheckPolicy', virtualGatewayHealthCheckPolicy_healthyThreshold - The number of consecutive successful health checks that must occur before declaring the listener healthy.

$sel:intervalMillis:VirtualGatewayHealthCheckPolicy', virtualGatewayHealthCheckPolicy_intervalMillis - The time period in milliseconds between each health check execution.

$sel:protocol:VirtualGatewayHealthCheckPolicy', virtualGatewayHealthCheckPolicy_protocol - The protocol for the health check request. If you specify grpc, then your service must conform to the GRPC Health Checking Protocol.

$sel:timeoutMillis:VirtualGatewayHealthCheckPolicy', virtualGatewayHealthCheckPolicy_timeoutMillis - The amount of time to wait when receiving a response from the health check, in milliseconds.

$sel:unhealthyThreshold:VirtualGatewayHealthCheckPolicy', virtualGatewayHealthCheckPolicy_unhealthyThreshold - The number of consecutive failed health checks that must occur before declaring a virtual gateway unhealthy.

virtualGatewayHealthCheckPolicy_path :: Lens' VirtualGatewayHealthCheckPolicy (Maybe Text)

The destination path for the health check request. This value is only used if the specified protocol is HTTP or HTTP/2. For any other protocol, this value is ignored.

virtualGatewayHealthCheckPolicy_port :: Lens' VirtualGatewayHealthCheckPolicy (Maybe Natural)

The destination port for the health check request. This port must match the port defined in the PortMapping for the listener.

virtualGatewayHealthCheckPolicy_healthyThreshold :: Lens' VirtualGatewayHealthCheckPolicy Natural

The number of consecutive successful health checks that must occur before declaring the listener healthy.

virtualGatewayHealthCheckPolicy_intervalMillis :: Lens' VirtualGatewayHealthCheckPolicy Natural

The time period in milliseconds between each health check execution.

virtualGatewayHealthCheckPolicy_protocol :: Lens' VirtualGatewayHealthCheckPolicy VirtualGatewayPortProtocol

The protocol for the health check request. If you specify grpc, then your service must conform to the GRPC Health Checking Protocol.

virtualGatewayHealthCheckPolicy_timeoutMillis :: Lens' VirtualGatewayHealthCheckPolicy Natural

The amount of time to wait when receiving a response from the health check, in milliseconds.

virtualGatewayHealthCheckPolicy_unhealthyThreshold :: Lens' VirtualGatewayHealthCheckPolicy Natural

The number of consecutive failed health checks that must occur before declaring a virtual gateway unhealthy.