Amazonka.AppMesh.Types.VirtualGatewayListenerTls

Description

Documentation

data VirtualGatewayListenerTls

An object that represents the Transport Layer Security (TLS) properties for a listener.

See: newVirtualGatewayListenerTls smart constructor.

Constructors

VirtualGatewayListenerTls'

Fields

validation :: Maybe VirtualGatewayListenerTlsValidationContext

A reference to an object that represents a virtual gateway's listener's Transport Layer Security (TLS) validation context.

certificate :: VirtualGatewayListenerTlsCertificate

An object that represents a Transport Layer Security (TLS) certificate.

mode :: VirtualGatewayListenerTlsMode

Specify one of the following modes.

____STRICT – Listener only accepts connections with TLS enabled.

____PERMISSIVE – Listener accepts connections with or without TLS enabled.

____DISABLED – Listener only accepts connections without TLS.

newVirtualGatewayListenerTls

Arguments

:: VirtualGatewayListenerTlsCertificate
$sel:certificate:VirtualGatewayListenerTls'

-> VirtualGatewayListenerTlsMode
$sel:mode:VirtualGatewayListenerTls'

-> VirtualGatewayListenerTls

Create a value of VirtualGatewayListenerTls with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:validation:VirtualGatewayListenerTls', virtualGatewayListenerTls_validation - A reference to an object that represents a virtual gateway's listener's Transport Layer Security (TLS) validation context.

$sel:certificate:VirtualGatewayListenerTls', virtualGatewayListenerTls_certificate - An object that represents a Transport Layer Security (TLS) certificate.

$sel:mode:VirtualGatewayListenerTls', virtualGatewayListenerTls_mode - Specify one of the following modes.

____STRICT – Listener only accepts connections with TLS enabled.

____PERMISSIVE – Listener accepts connections with or without TLS enabled.

____DISABLED – Listener only accepts connections without TLS.

virtualGatewayListenerTls_validation :: Lens' VirtualGatewayListenerTls (Maybe VirtualGatewayListenerTlsValidationContext)

A reference to an object that represents a virtual gateway's listener's Transport Layer Security (TLS) validation context.

virtualGatewayListenerTls_certificate :: Lens' VirtualGatewayListenerTls VirtualGatewayListenerTlsCertificate

An object that represents a Transport Layer Security (TLS) certificate.

virtualGatewayListenerTls_mode :: Lens' VirtualGatewayListenerTls VirtualGatewayListenerTlsMode

Specify one of the following modes.

____STRICT – Listener only accepts connections with TLS enabled.

____PERMISSIVE – Listener accepts connections with or without TLS enabled.

____DISABLED – Listener only accepts connections without TLS.