Amazonka.AppMesh.Types.VirtualGatewaySpec

Description

Documentation

data VirtualGatewaySpec

An object that represents the specification of a service mesh resource.

See: newVirtualGatewaySpec smart constructor.

Constructors

VirtualGatewaySpec'

Fields

backendDefaults :: Maybe VirtualGatewayBackendDefaults

A reference to an object that represents the defaults for backends.

logging :: Maybe VirtualGatewayLogging

listeners :: [VirtualGatewayListener]

The listeners that the mesh endpoint is expected to receive inbound traffic from. You can specify one listener.

newVirtualGatewaySpec :: VirtualGatewaySpec

Create a value of VirtualGatewaySpec with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:backendDefaults:VirtualGatewaySpec', virtualGatewaySpec_backendDefaults - A reference to an object that represents the defaults for backends.

$sel:logging:VirtualGatewaySpec', virtualGatewaySpec_logging - Undocumented member.

$sel:listeners:VirtualGatewaySpec', virtualGatewaySpec_listeners - The listeners that the mesh endpoint is expected to receive inbound traffic from. You can specify one listener.

virtualGatewaySpec_backendDefaults :: Lens' VirtualGatewaySpec (Maybe VirtualGatewayBackendDefaults)

A reference to an object that represents the defaults for backends.

virtualGatewaySpec_logging :: Lens' VirtualGatewaySpec (Maybe VirtualGatewayLogging)

Undocumented member.

virtualGatewaySpec_listeners :: Lens' VirtualGatewaySpec [VirtualGatewayListener]

The listeners that the mesh endpoint is expected to receive inbound traffic from. You can specify one listener.