Amazonka.AppMesh.Types.VirtualNodeConnectionPool

Description

Documentation

data VirtualNodeConnectionPool

An object that represents the type of virtual node connection pool.

Only one protocol is used at a time and should be the same protocol as the one chosen under port mapping.

If not present the default value for maxPendingRequests is 2147483647.

See: newVirtualNodeConnectionPool smart constructor.

Constructors

VirtualNodeConnectionPool'

Fields

http2 :: Maybe VirtualNodeHttp2ConnectionPool

An object that represents a type of connection pool.

grpc :: Maybe VirtualNodeGrpcConnectionPool

An object that represents a type of connection pool.

tcp :: Maybe VirtualNodeTcpConnectionPool

An object that represents a type of connection pool.

http :: Maybe VirtualNodeHttpConnectionPool

An object that represents a type of connection pool.

newVirtualNodeConnectionPool :: VirtualNodeConnectionPool

Create a value of VirtualNodeConnectionPool with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:http2:VirtualNodeConnectionPool', virtualNodeConnectionPool_http2 - An object that represents a type of connection pool.

$sel:grpc:VirtualNodeConnectionPool', virtualNodeConnectionPool_grpc - An object that represents a type of connection pool.

$sel:tcp:VirtualNodeConnectionPool', virtualNodeConnectionPool_tcp - An object that represents a type of connection pool.

$sel:http:VirtualNodeConnectionPool', virtualNodeConnectionPool_http - An object that represents a type of connection pool.

virtualNodeConnectionPool_http2 :: Lens' VirtualNodeConnectionPool (Maybe VirtualNodeHttp2ConnectionPool)

An object that represents a type of connection pool.

virtualNodeConnectionPool_grpc :: Lens' VirtualNodeConnectionPool (Maybe VirtualNodeGrpcConnectionPool)

An object that represents a type of connection pool.

virtualNodeConnectionPool_tcp :: Lens' VirtualNodeConnectionPool (Maybe VirtualNodeTcpConnectionPool)

An object that represents a type of connection pool.

virtualNodeConnectionPool_http :: Lens' VirtualNodeConnectionPool (Maybe VirtualNodeHttpConnectionPool)

An object that represents a type of connection pool.