Amazonka.AppMesh.Types.VirtualNodeData

Description

Documentation

data VirtualNodeData

An object that represents a virtual node returned by a describe operation.

See: newVirtualNodeData smart constructor.

Constructors

VirtualNodeData'

Fields

meshName :: Text

The name of the service mesh that the virtual node resides in.

metadata :: ResourceMetadata

The associated metadata for the virtual node.

spec :: VirtualNodeSpec

The specifications of the virtual node.

status :: VirtualNodeStatus

The current status for the virtual node.

virtualNodeName :: Text

The name of the virtual node.

newVirtualNodeData

Arguments

:: Text
$sel:meshName:VirtualNodeData'

-> ResourceMetadata
$sel:metadata:VirtualNodeData'

-> VirtualNodeSpec
$sel:spec:VirtualNodeData'

-> VirtualNodeStatus
$sel:status:VirtualNodeData'

-> Text
$sel:virtualNodeName:VirtualNodeData'

-> VirtualNodeData

Create a value of VirtualNodeData with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:meshName:VirtualNodeData', virtualNodeData_meshName - The name of the service mesh that the virtual node resides in.

$sel:metadata:VirtualNodeData', virtualNodeData_metadata - The associated metadata for the virtual node.

$sel:spec:VirtualNodeData', virtualNodeData_spec - The specifications of the virtual node.

$sel:status:VirtualNodeData', virtualNodeData_status - The current status for the virtual node.

$sel:virtualNodeName:VirtualNodeData', virtualNodeData_virtualNodeName - The name of the virtual node.

virtualNodeData_meshName :: Lens' VirtualNodeData Text

The name of the service mesh that the virtual node resides in.

virtualNodeData_metadata :: Lens' VirtualNodeData ResourceMetadata

The associated metadata for the virtual node.

virtualNodeData_spec :: Lens' VirtualNodeData VirtualNodeSpec

The specifications of the virtual node.

virtualNodeData_status :: Lens' VirtualNodeData VirtualNodeStatus

The current status for the virtual node.

virtualNodeData_virtualNodeName :: Lens' VirtualNodeData Text

The name of the virtual node.