Amazonka.AppMesh.Types.VirtualNodeSpec

Description

Documentation

data VirtualNodeSpec

An object that represents the specification of a virtual node.

See: newVirtualNodeSpec smart constructor.

Constructors

VirtualNodeSpec'

Fields

backends :: Maybe [Backend]

The backends that the virtual node is expected to send outbound traffic to.

backendDefaults :: Maybe BackendDefaults

A reference to an object that represents the defaults for backends.

serviceDiscovery :: Maybe ServiceDiscovery

The service discovery information for the virtual node. If your virtual node does not expect ingress traffic, you can omit this parameter. If you specify a listener, then you must specify service discovery information.

listeners :: Maybe [Listener]

The listener that the virtual node is expected to receive inbound traffic from. You can specify one listener.

logging :: Maybe Logging

The inbound and outbound access logging information for the virtual node.

newVirtualNodeSpec :: VirtualNodeSpec

Create a value of VirtualNodeSpec with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:backends:VirtualNodeSpec', virtualNodeSpec_backends - The backends that the virtual node is expected to send outbound traffic to.

$sel:backendDefaults:VirtualNodeSpec', virtualNodeSpec_backendDefaults - A reference to an object that represents the defaults for backends.

$sel:serviceDiscovery:VirtualNodeSpec', virtualNodeSpec_serviceDiscovery - The service discovery information for the virtual node. If your virtual node does not expect ingress traffic, you can omit this parameter. If you specify a listener, then you must specify service discovery information.

$sel:listeners:VirtualNodeSpec', virtualNodeSpec_listeners - The listener that the virtual node is expected to receive inbound traffic from. You can specify one listener.

$sel:logging:VirtualNodeSpec', virtualNodeSpec_logging - The inbound and outbound access logging information for the virtual node.

virtualNodeSpec_backends :: Lens' VirtualNodeSpec (Maybe [Backend])

The backends that the virtual node is expected to send outbound traffic to.

virtualNodeSpec_backendDefaults :: Lens' VirtualNodeSpec (Maybe BackendDefaults)

A reference to an object that represents the defaults for backends.

virtualNodeSpec_serviceDiscovery :: Lens' VirtualNodeSpec (Maybe ServiceDiscovery)

The service discovery information for the virtual node. If your virtual node does not expect ingress traffic, you can omit this parameter. If you specify a listener, then you must specify service discovery information.

virtualNodeSpec_listeners :: Lens' VirtualNodeSpec (Maybe [Listener])

The listener that the virtual node is expected to receive inbound traffic from. You can specify one listener.

virtualNodeSpec_logging :: Lens' VirtualNodeSpec (Maybe Logging)

The inbound and outbound access logging information for the virtual node.