Amazonka.AppMesh.UpdateGatewayRoute

Description

Updates an existing gateway route that is associated to a specified virtual gateway in a service mesh.

Creating a Request

data UpdateGatewayRoute

See: newUpdateGatewayRoute smart constructor.

Constructors

UpdateGatewayRoute'

Fields

clientToken :: Maybe Text

Unique, case-sensitive identifier that you provide to ensure the idempotency of the request. Up to 36 letters, numbers, hyphens, and underscores are allowed.

meshOwner :: Maybe Text

The AWS IAM account ID of the service mesh owner. If the account ID is not your own, then it's the ID of the account that shared the mesh with your account. For more information about mesh sharing, see Working with shared meshes.

gatewayRouteName :: Text

The name of the gateway route to update.

meshName :: Text

The name of the service mesh that the gateway route resides in.

spec :: GatewayRouteSpec

The new gateway route specification to apply. This overwrites the existing data.

virtualGatewayName :: Text

The name of the virtual gateway that the gateway route is associated with.

newUpdateGatewayRoute

Arguments

:: Text
$sel:gatewayRouteName:UpdateGatewayRoute'

-> Text
$sel:meshName:UpdateGatewayRoute'

-> GatewayRouteSpec
$sel:spec:UpdateGatewayRoute'

-> Text
$sel:virtualGatewayName:UpdateGatewayRoute'

-> UpdateGatewayRoute

Create a value of UpdateGatewayRoute with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:clientToken:UpdateGatewayRoute', updateGatewayRoute_clientToken - Unique, case-sensitive identifier that you provide to ensure the idempotency of the request. Up to 36 letters, numbers, hyphens, and underscores are allowed.

$sel:meshOwner:UpdateGatewayRoute', updateGatewayRoute_meshOwner - The AWS IAM account ID of the service mesh owner. If the account ID is not your own, then it's the ID of the account that shared the mesh with your account. For more information about mesh sharing, see Working with shared meshes.

$sel:gatewayRouteName:UpdateGatewayRoute', updateGatewayRoute_gatewayRouteName - The name of the gateway route to update.

$sel:meshName:UpdateGatewayRoute', updateGatewayRoute_meshName - The name of the service mesh that the gateway route resides in.

$sel:spec:UpdateGatewayRoute', updateGatewayRoute_spec - The new gateway route specification to apply. This overwrites the existing data.

$sel:virtualGatewayName:UpdateGatewayRoute', updateGatewayRoute_virtualGatewayName - The name of the virtual gateway that the gateway route is associated with.

Request Lenses

updateGatewayRoute_clientToken :: Lens' UpdateGatewayRoute (Maybe Text)

Unique, case-sensitive identifier that you provide to ensure the idempotency of the request. Up to 36 letters, numbers, hyphens, and underscores are allowed.

updateGatewayRoute_meshOwner :: Lens' UpdateGatewayRoute (Maybe Text)

The AWS IAM account ID of the service mesh owner. If the account ID is not your own, then it's the ID of the account that shared the mesh with your account. For more information about mesh sharing, see Working with shared meshes.

updateGatewayRoute_gatewayRouteName :: Lens' UpdateGatewayRoute Text

The name of the gateway route to update.

updateGatewayRoute_meshName :: Lens' UpdateGatewayRoute Text

The name of the service mesh that the gateway route resides in.

updateGatewayRoute_spec :: Lens' UpdateGatewayRoute GatewayRouteSpec

The new gateway route specification to apply. This overwrites the existing data.

updateGatewayRoute_virtualGatewayName :: Lens' UpdateGatewayRoute Text

The name of the virtual gateway that the gateway route is associated with.

Destructuring the Response

data UpdateGatewayRouteResponse

See: newUpdateGatewayRouteResponse smart constructor.

Constructors

UpdateGatewayRouteResponse'

Fields

httpStatus :: Int

The response's http status code.

gatewayRoute :: GatewayRouteData

A full description of the gateway route that was updated.

newUpdateGatewayRouteResponse

Arguments

:: Int
$sel:httpStatus:UpdateGatewayRouteResponse'

-> GatewayRouteData
$sel:gatewayRoute:UpdateGatewayRouteResponse'

-> UpdateGatewayRouteResponse

Create a value of UpdateGatewayRouteResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:httpStatus:UpdateGatewayRouteResponse', updateGatewayRouteResponse_httpStatus - The response's http status code.

$sel:gatewayRoute:UpdateGatewayRouteResponse', updateGatewayRouteResponse_gatewayRoute - A full description of the gateway route that was updated.

Response Lenses

updateGatewayRouteResponse_httpStatus :: Lens' UpdateGatewayRouteResponse Int

The response's http status code.

updateGatewayRouteResponse_gatewayRoute :: Lens' UpdateGatewayRouteResponse GatewayRouteData

A full description of the gateway route that was updated.