Amazonka.AppMesh.UpdateVirtualService

Description

Updates an existing virtual service in a specified service mesh.

Creating a Request

data UpdateVirtualService

See: newUpdateVirtualService smart constructor.

Constructors

UpdateVirtualService'

Fields

clientToken :: Maybe Text

Unique, case-sensitive identifier that you provide to ensure the idempotency of the request. Up to 36 letters, numbers, hyphens, and underscores are allowed.

meshOwner :: Maybe Text

The AWS IAM account ID of the service mesh owner. If the account ID is not your own, then it's the ID of the account that shared the mesh with your account. For more information about mesh sharing, see Working with shared meshes.

meshName :: Text

The name of the service mesh that the virtual service resides in.

spec :: VirtualServiceSpec

The new virtual service specification to apply. This overwrites the existing data.

virtualServiceName :: Text

The name of the virtual service to update.

newUpdateVirtualService

Arguments

:: Text
$sel:meshName:UpdateVirtualService'

-> VirtualServiceSpec
$sel:spec:UpdateVirtualService'

-> Text
$sel:virtualServiceName:UpdateVirtualService'

-> UpdateVirtualService

Create a value of UpdateVirtualService with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:clientToken:UpdateVirtualService', updateVirtualService_clientToken - Unique, case-sensitive identifier that you provide to ensure the idempotency of the request. Up to 36 letters, numbers, hyphens, and underscores are allowed.

$sel:meshOwner:UpdateVirtualService', updateVirtualService_meshOwner - The AWS IAM account ID of the service mesh owner. If the account ID is not your own, then it's the ID of the account that shared the mesh with your account. For more information about mesh sharing, see Working with shared meshes.

$sel:meshName:UpdateVirtualService', updateVirtualService_meshName - The name of the service mesh that the virtual service resides in.

$sel:spec:UpdateVirtualService', updateVirtualService_spec - The new virtual service specification to apply. This overwrites the existing data.

$sel:virtualServiceName:UpdateVirtualService', updateVirtualService_virtualServiceName - The name of the virtual service to update.

Request Lenses

updateVirtualService_clientToken :: Lens' UpdateVirtualService (Maybe Text)

Unique, case-sensitive identifier that you provide to ensure the idempotency of the request. Up to 36 letters, numbers, hyphens, and underscores are allowed.

updateVirtualService_meshOwner :: Lens' UpdateVirtualService (Maybe Text)

The AWS IAM account ID of the service mesh owner. If the account ID is not your own, then it's the ID of the account that shared the mesh with your account. For more information about mesh sharing, see Working with shared meshes.

updateVirtualService_meshName :: Lens' UpdateVirtualService Text

The name of the service mesh that the virtual service resides in.

updateVirtualService_spec :: Lens' UpdateVirtualService VirtualServiceSpec

The new virtual service specification to apply. This overwrites the existing data.

updateVirtualService_virtualServiceName :: Lens' UpdateVirtualService Text

The name of the virtual service to update.

Destructuring the Response

data UpdateVirtualServiceResponse

See: newUpdateVirtualServiceResponse smart constructor.

Constructors

UpdateVirtualServiceResponse'

Fields

httpStatus :: Int

The response's http status code.

virtualService :: VirtualServiceData

A full description of the virtual service that was updated.

newUpdateVirtualServiceResponse

Arguments

:: Int
$sel:httpStatus:UpdateVirtualServiceResponse'

-> VirtualServiceData
$sel:virtualService:UpdateVirtualServiceResponse'

-> UpdateVirtualServiceResponse

Create a value of UpdateVirtualServiceResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:httpStatus:UpdateVirtualServiceResponse', updateVirtualServiceResponse_httpStatus - The response's http status code.

$sel:virtualService:UpdateVirtualServiceResponse', updateVirtualServiceResponse_virtualService - A full description of the virtual service that was updated.

Response Lenses

updateVirtualServiceResponse_httpStatus :: Lens' UpdateVirtualServiceResponse Int

The response's http status code.

updateVirtualServiceResponse_virtualService :: Lens' UpdateVirtualServiceResponse VirtualServiceData

A full description of the virtual service that was updated.