Amazonka.AppRunner.CreateAutoScalingConfiguration

Description

Create an App Runner automatic scaling configuration resource. App Runner requires this resource when you create App Runner services that require non-default auto scaling settings. You can share an auto scaling configuration across multiple services.

Create multiple revisions of a configuration by using the same AutoScalingConfigurationName and different AutoScalingConfigurationRevision values. When you create a service, you can set it to use the latest active revision of an auto scaling configuration or a specific revision.

Configure a higher MinSize to increase the spread of your App Runner service over more Availability Zones in the Amazon Web Services Region. The tradeoff is a higher minimal cost.

Configure a lower MaxSize to control your cost. The tradeoff is lower responsiveness during peak demand.

Creating a Request

data CreateAutoScalingConfiguration

See: newCreateAutoScalingConfiguration smart constructor.

Constructors

CreateAutoScalingConfiguration'

Fields

maxSize :: Maybe Natural

The maximum number of instances that your service scales up to. At most MaxSize instances actively serve traffic for your service.

Default: 25

minSize :: Maybe Natural

The minimum number of instances that App Runner provisions for your service. The service always has at least MinSize provisioned instances. Some of them actively serve traffic. The rest of them (provisioned and inactive instances) are a cost-effective compute capacity reserve and are ready to be quickly activated. You pay for memory usage of all the provisioned instances. You pay for CPU usage of only the active subset.

App Runner temporarily doubles the number of provisioned instances during deployments, to maintain the same capacity for both old and new code.

Default: 1

tags :: Maybe [Tag]

A list of metadata items that you can associate with your auto scaling configuration resource. A tag is a key-value pair.

maxConcurrency :: Maybe Natural

The maximum number of concurrent requests that you want an instance to process. If the number of concurrent requests exceeds this limit, App Runner scales up your service.

Default: 100

autoScalingConfigurationName :: Text

A name for the auto scaling configuration. When you use it for the first time in an Amazon Web Services Region, App Runner creates revision number 1 of this name. When you use the same name in subsequent calls, App Runner creates incremental revisions of the configuration.

newCreateAutoScalingConfiguration

Arguments

:: Text
$sel:autoScalingConfigurationName:CreateAutoScalingConfiguration'

-> CreateAutoScalingConfiguration

Create a value of CreateAutoScalingConfiguration with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:maxSize:CreateAutoScalingConfiguration', createAutoScalingConfiguration_maxSize - The maximum number of instances that your service scales up to. At most MaxSize instances actively serve traffic for your service.

Default: 25

$sel:minSize:CreateAutoScalingConfiguration', createAutoScalingConfiguration_minSize - The minimum number of instances that App Runner provisions for your service. The service always has at least MinSize provisioned instances. Some of them actively serve traffic. The rest of them (provisioned and inactive instances) are a cost-effective compute capacity reserve and are ready to be quickly activated. You pay for memory usage of all the provisioned instances. You pay for CPU usage of only the active subset.

App Runner temporarily doubles the number of provisioned instances during deployments, to maintain the same capacity for both old and new code.

Default: 1

$sel:tags:CreateAutoScalingConfiguration', createAutoScalingConfiguration_tags - A list of metadata items that you can associate with your auto scaling configuration resource. A tag is a key-value pair.

$sel:maxConcurrency:CreateAutoScalingConfiguration', createAutoScalingConfiguration_maxConcurrency - The maximum number of concurrent requests that you want an instance to process. If the number of concurrent requests exceeds this limit, App Runner scales up your service.

Default: 100

$sel:autoScalingConfigurationName:CreateAutoScalingConfiguration', createAutoScalingConfiguration_autoScalingConfigurationName - A name for the auto scaling configuration. When you use it for the first time in an Amazon Web Services Region, App Runner creates revision number 1 of this name. When you use the same name in subsequent calls, App Runner creates incremental revisions of the configuration.

Request Lenses

createAutoScalingConfiguration_maxSize :: Lens' CreateAutoScalingConfiguration (Maybe Natural)

The maximum number of instances that your service scales up to. At most MaxSize instances actively serve traffic for your service.

Default: 25

createAutoScalingConfiguration_minSize :: Lens' CreateAutoScalingConfiguration (Maybe Natural)

The minimum number of instances that App Runner provisions for your service. The service always has at least MinSize provisioned instances. Some of them actively serve traffic. The rest of them (provisioned and inactive instances) are a cost-effective compute capacity reserve and are ready to be quickly activated. You pay for memory usage of all the provisioned instances. You pay for CPU usage of only the active subset.

App Runner temporarily doubles the number of provisioned instances during deployments, to maintain the same capacity for both old and new code.

Default: 1

createAutoScalingConfiguration_tags :: Lens' CreateAutoScalingConfiguration (Maybe [Tag])

A list of metadata items that you can associate with your auto scaling configuration resource. A tag is a key-value pair.

createAutoScalingConfiguration_maxConcurrency :: Lens' CreateAutoScalingConfiguration (Maybe Natural)

The maximum number of concurrent requests that you want an instance to process. If the number of concurrent requests exceeds this limit, App Runner scales up your service.

Default: 100

createAutoScalingConfiguration_autoScalingConfigurationName :: Lens' CreateAutoScalingConfiguration Text

A name for the auto scaling configuration. When you use it for the first time in an Amazon Web Services Region, App Runner creates revision number 1 of this name. When you use the same name in subsequent calls, App Runner creates incremental revisions of the configuration.

Destructuring the Response

data CreateAutoScalingConfigurationResponse

See: newCreateAutoScalingConfigurationResponse smart constructor.

Constructors

CreateAutoScalingConfigurationResponse'

Fields

httpStatus :: Int

The response's http status code.

autoScalingConfiguration :: AutoScalingConfiguration

A description of the App Runner auto scaling configuration that's created by this request.

newCreateAutoScalingConfigurationResponse

Arguments

:: Int
$sel:httpStatus:CreateAutoScalingConfigurationResponse'

-> AutoScalingConfiguration
$sel:autoScalingConfiguration:CreateAutoScalingConfigurationResponse'

-> CreateAutoScalingConfigurationResponse

Create a value of CreateAutoScalingConfigurationResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:httpStatus:CreateAutoScalingConfigurationResponse', createAutoScalingConfigurationResponse_httpStatus - The response's http status code.

$sel:autoScalingConfiguration:CreateAutoScalingConfigurationResponse', createAutoScalingConfigurationResponse_autoScalingConfiguration - A description of the App Runner auto scaling configuration that's created by this request.

Response Lenses

createAutoScalingConfigurationResponse_httpStatus :: Lens' CreateAutoScalingConfigurationResponse Int

The response's http status code.

createAutoScalingConfigurationResponse_autoScalingConfiguration :: Lens' CreateAutoScalingConfigurationResponse AutoScalingConfiguration

A description of the App Runner auto scaling configuration that's created by this request.