Amazonka.AppRunner.CreateService

Description

Create an App Runner service. After the service is created, the action also automatically starts a deployment.

This is an asynchronous operation. On a successful call, you can use the returned OperationId and the ListOperations call to track the operation's progress.

Creating a Request

data CreateService

See: newCreateService smart constructor.

Constructors

CreateService'

Fields

autoScalingConfigurationArn :: Maybe Text

The Amazon Resource Name (ARN) of an App Runner automatic scaling configuration resource that you want to associate with your service. If not provided, App Runner associates the latest revision of a default auto scaling configuration.

encryptionConfiguration :: Maybe EncryptionConfiguration

An optional custom encryption key that App Runner uses to encrypt the copy of your source repository that it maintains and your service logs. By default, App Runner uses an Amazon Web Services managed CMK.

healthCheckConfiguration :: Maybe HealthCheckConfiguration

The settings for the health check that App Runner performs to monitor the health of your service.

tags :: Maybe [Tag]

An optional list of metadata items that you can associate with your service resource. A tag is a key-value pair.

instanceConfiguration :: Maybe InstanceConfiguration

The runtime configuration of instances (scaling units) of the App Runner service.

serviceName :: Text

A name for the new service. It must be unique across all the running App Runner services in your Amazon Web Services account in the Amazon Web Services Region.

sourceConfiguration :: SourceConfiguration

The source to deploy to the App Runner service. It can be a code or an image repository.

newCreateService

Arguments

:: Text
$sel:serviceName:CreateService'

-> SourceConfiguration
$sel:sourceConfiguration:CreateService'

-> CreateService

Create a value of CreateService with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:autoScalingConfigurationArn:CreateService', createService_autoScalingConfigurationArn - The Amazon Resource Name (ARN) of an App Runner automatic scaling configuration resource that you want to associate with your service. If not provided, App Runner associates the latest revision of a default auto scaling configuration.

$sel:encryptionConfiguration:CreateService', createService_encryptionConfiguration - An optional custom encryption key that App Runner uses to encrypt the copy of your source repository that it maintains and your service logs. By default, App Runner uses an Amazon Web Services managed CMK.

$sel:healthCheckConfiguration:CreateService', createService_healthCheckConfiguration - The settings for the health check that App Runner performs to monitor the health of your service.

$sel:tags:CreateService', createService_tags - An optional list of metadata items that you can associate with your service resource. A tag is a key-value pair.

$sel:instanceConfiguration:CreateService', createService_instanceConfiguration - The runtime configuration of instances (scaling units) of the App Runner service.

$sel:serviceName:CreateService', createService_serviceName - A name for the new service. It must be unique across all the running App Runner services in your Amazon Web Services account in the Amazon Web Services Region.

$sel:sourceConfiguration:CreateService', createService_sourceConfiguration - The source to deploy to the App Runner service. It can be a code or an image repository.

Request Lenses

createService_autoScalingConfigurationArn :: Lens' CreateService (Maybe Text)

The Amazon Resource Name (ARN) of an App Runner automatic scaling configuration resource that you want to associate with your service. If not provided, App Runner associates the latest revision of a default auto scaling configuration.

createService_encryptionConfiguration :: Lens' CreateService (Maybe EncryptionConfiguration)

An optional custom encryption key that App Runner uses to encrypt the copy of your source repository that it maintains and your service logs. By default, App Runner uses an Amazon Web Services managed CMK.

createService_healthCheckConfiguration :: Lens' CreateService (Maybe HealthCheckConfiguration)

The settings for the health check that App Runner performs to monitor the health of your service.

createService_tags :: Lens' CreateService (Maybe [Tag])

An optional list of metadata items that you can associate with your service resource. A tag is a key-value pair.

createService_instanceConfiguration :: Lens' CreateService (Maybe InstanceConfiguration)

The runtime configuration of instances (scaling units) of the App Runner service.

createService_serviceName :: Lens' CreateService Text

A name for the new service. It must be unique across all the running App Runner services in your Amazon Web Services account in the Amazon Web Services Region.

createService_sourceConfiguration :: Lens' CreateService SourceConfiguration

The source to deploy to the App Runner service. It can be a code or an image repository.

Destructuring the Response

data CreateServiceResponse

See: newCreateServiceResponse smart constructor.

Constructors

CreateServiceResponse'

Fields

httpStatus :: Int

The response's http status code.

service :: Service

A description of the App Runner service that's created by this request.

operationId :: Text

The unique ID of the asynchronous operation that this request started. You can use it combined with the ListOperations call to track the operation's progress.

newCreateServiceResponse

Arguments

:: Int
$sel:httpStatus:CreateServiceResponse'

-> Service
$sel:service:CreateServiceResponse'

-> Text
$sel:operationId:CreateServiceResponse'

-> CreateServiceResponse

Create a value of CreateServiceResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:httpStatus:CreateServiceResponse', createServiceResponse_httpStatus - The response's http status code.

$sel:service:CreateServiceResponse', createServiceResponse_service - A description of the App Runner service that's created by this request.

$sel:operationId:CreateServiceResponse', createServiceResponse_operationId - The unique ID of the asynchronous operation that this request started. You can use it combined with the ListOperations call to track the operation's progress.

Response Lenses

createServiceResponse_httpStatus :: Lens' CreateServiceResponse Int

The response's http status code.

createServiceResponse_service :: Lens' CreateServiceResponse Service

A description of the App Runner service that's created by this request.

createServiceResponse_operationId :: Lens' CreateServiceResponse Text

The unique ID of the asynchronous operation that this request started. You can use it combined with the ListOperations call to track the operation's progress.