Amazonka.AppRunner.Lens

Description

Operations

ListServices

listServices_nextToken :: Lens' ListServices (Maybe Text)

A token from a previous result page. Used for a paginated request. The request retrieves the next result page. All other parameter values must be identical to the ones specified in the initial request.

If you don't specify NextToken, the request retrieves the first result page.

listServices_maxResults :: Lens' ListServices (Maybe Natural)

The maximum number of results to include in each response (result page). It's used for a paginated request.

If you don't specify MaxResults, the request retrieves all available results in a single response.

listServicesResponse_nextToken :: Lens' ListServicesResponse (Maybe Text)

The token that you can pass in a subsequent request to get the next result page. It's returned in a paginated request.

listServicesResponse_httpStatus :: Lens' ListServicesResponse Int

The response's http status code.

listServicesResponse_serviceSummaryList :: Lens' ListServicesResponse [ServiceSummary]

A list of service summary information records. In a paginated request, the request returns up to MaxResults records for each call.

ResumeService

resumeService_serviceArn :: Lens' ResumeService Text

The Amazon Resource Name (ARN) of the App Runner service that you want to resume.

resumeServiceResponse_operationId :: Lens' ResumeServiceResponse (Maybe Text)

The unique ID of the asynchronous operation that this request started. You can use it combined with the ListOperations call to track the operation's progress.

resumeServiceResponse_httpStatus :: Lens' ResumeServiceResponse Int

The response's http status code.

resumeServiceResponse_service :: Lens' ResumeServiceResponse Service

A description of the App Runner service that this request just resumed.

DeleteService

deleteService_serviceArn :: Lens' DeleteService Text

The Amazon Resource Name (ARN) of the App Runner service that you want to delete.

deleteServiceResponse_httpStatus :: Lens' DeleteServiceResponse Int

The response's http status code.

deleteServiceResponse_service :: Lens' DeleteServiceResponse Service

A description of the App Runner service that this request just deleted.

deleteServiceResponse_operationId :: Lens' DeleteServiceResponse Text

The unique ID of the asynchronous operation that this request started. You can use it combined with the ListOperations call to track the operation's progress.

UpdateService

updateService_autoScalingConfigurationArn :: Lens' UpdateService (Maybe Text)

The Amazon Resource Name (ARN) of an App Runner automatic scaling configuration resource that you want to associate with your service.

updateService_healthCheckConfiguration :: Lens' UpdateService (Maybe HealthCheckConfiguration)

The settings for the health check that App Runner performs to monitor the health of your service.

updateService_sourceConfiguration :: Lens' UpdateService (Maybe SourceConfiguration)

The source configuration to apply to the App Runner service.

You can change the configuration of the code or image repository that the service uses. However, you can't switch from code to image or the other way around. This means that you must provide the same structure member of SourceConfiguration that you originally included when you created the service. Specifically, you can include either CodeRepository or ImageRepository. To update the source configuration, set the values to members of the structure that you include.

updateService_instanceConfiguration :: Lens' UpdateService (Maybe InstanceConfiguration)

The runtime configuration to apply to instances (scaling units) of the App Runner service.

updateService_serviceArn :: Lens' UpdateService Text

The Amazon Resource Name (ARN) of the App Runner service that you want to update.

updateServiceResponse_httpStatus :: Lens' UpdateServiceResponse Int

The response's http status code.

updateServiceResponse_service :: Lens' UpdateServiceResponse Service

A description of the App Runner service updated by this request. All configuration values in the returned Service structure reflect configuration changes that are being applied by this request.

updateServiceResponse_operationId :: Lens' UpdateServiceResponse Text

The unique ID of the asynchronous operation that this request started. You can use it combined with the ListOperations call to track the operation's progress.

ListOperations

listOperations_nextToken :: Lens' ListOperations (Maybe Text)

A token from a previous result page. It's used for a paginated request. The request retrieves the next result page. All other parameter values must be identical to the ones specified in the initial request.

If you don't specify NextToken, the request retrieves the first result page.

listOperations_maxResults :: Lens' ListOperations (Maybe Natural)

The maximum number of results to include in each response (result page). It's used for a paginated request.

If you don't specify MaxResults, the request retrieves all available results in a single response.

listOperations_serviceArn :: Lens' ListOperations Text

The Amazon Resource Name (ARN) of the App Runner service that you want a list of operations for.

listOperationsResponse_nextToken :: Lens' ListOperationsResponse (Maybe Text)

The token that you can pass in a subsequent request to get the next result page. It's returned in a paginated request.

listOperationsResponse_operationSummaryList :: Lens' ListOperationsResponse (Maybe [OperationSummary])

A list of operation summary information records. In a paginated request, the request returns up to MaxResults records for each call.

listOperationsResponse_httpStatus :: Lens' ListOperationsResponse Int

The response's http status code.

AssociateCustomDomain

associateCustomDomain_enableWWWSubdomain :: Lens' AssociateCustomDomain (Maybe Bool)

Set to true to associate the subdomain www.DomainName with the App Runner service in addition to the base domain.

Default: true

associateCustomDomain_serviceArn :: Lens' AssociateCustomDomain Text

The Amazon Resource Name (ARN) of the App Runner service that you want to associate a custom domain name with.

associateCustomDomain_domainName :: Lens' AssociateCustomDomain Text

A custom domain endpoint to associate. Specify a root domain (for example, example.com), a subdomain (for example, login.example.com or admin.login.example.com), or a wildcard (for example, *.example.com).

associateCustomDomainResponse_httpStatus :: Lens' AssociateCustomDomainResponse Int

The response's http status code.

associateCustomDomainResponse_dNSTarget :: Lens' AssociateCustomDomainResponse Text

The App Runner subdomain of the App Runner service. The custom domain name is mapped to this target name.

associateCustomDomainResponse_serviceArn :: Lens' AssociateCustomDomainResponse Text

The Amazon Resource Name (ARN) of the App Runner service with which a custom domain name is associated.

associateCustomDomainResponse_customDomain :: Lens' AssociateCustomDomainResponse CustomDomain

A description of the domain name that's being associated.

ListConnections

listConnections_connectionName :: Lens' ListConnections (Maybe Text)

If specified, only this connection is returned. If not specified, the result isn't filtered by name.

listConnections_nextToken :: Lens' ListConnections (Maybe Text)

A token from a previous result page. Used for a paginated request. The request retrieves the next result page. All other parameter values must be identical to the ones specified in the initial request.

If you don't specify NextToken, the request retrieves the first result page.

listConnections_maxResults :: Lens' ListConnections (Maybe Natural)

The maximum number of results to include in each response (result page). Used for a paginated request.

If you don't specify MaxResults, the request retrieves all available results in a single response.

listConnectionsResponse_nextToken :: Lens' ListConnectionsResponse (Maybe Text)

The token that you can pass in a subsequent request to get the next result page. Returned in a paginated request.

listConnectionsResponse_httpStatus :: Lens' ListConnectionsResponse Int

The response's http status code.

listConnectionsResponse_connectionSummaryList :: Lens' ListConnectionsResponse [ConnectionSummary]

A list of summary information records for connections. In a paginated request, the request returns up to MaxResults records for each call.

DeleteConnection

deleteConnection_connectionArn :: Lens' DeleteConnection Text

The Amazon Resource Name (ARN) of the App Runner connection that you want to delete.

deleteConnectionResponse_connection :: Lens' DeleteConnectionResponse (Maybe Connection)

A description of the App Runner connection that this request just deleted.

deleteConnectionResponse_httpStatus :: Lens' DeleteConnectionResponse Int

The response's http status code.

DescribeAutoScalingConfiguration

describeAutoScalingConfiguration_autoScalingConfigurationArn :: Lens' DescribeAutoScalingConfiguration Text

The Amazon Resource Name (ARN) of the App Runner auto scaling configuration that you want a description for.

The ARN can be a full auto scaling configuration ARN, or a partial ARN ending with either .../name or .../name/revision . If a revision isn't specified, the latest active revision is described.

describeAutoScalingConfigurationResponse_httpStatus :: Lens' DescribeAutoScalingConfigurationResponse Int

The response's http status code.

describeAutoScalingConfigurationResponse_autoScalingConfiguration :: Lens' DescribeAutoScalingConfigurationResponse AutoScalingConfiguration

A full description of the App Runner auto scaling configuration that you specified in this request.

ListTagsForResource

listTagsForResource_resourceArn :: Lens' ListTagsForResource Text

The Amazon Resource Name (ARN) of the resource that a tag list is requested for.

It must be the ARN of an App Runner resource.

listTagsForResourceResponse_tags :: Lens' ListTagsForResourceResponse (Maybe [Tag])

A list of the tag key-value pairs that are associated with the resource.

listTagsForResourceResponse_httpStatus :: Lens' ListTagsForResourceResponse Int

The response's http status code.

CreateConnection

createConnection_tags :: Lens' CreateConnection (Maybe [Tag])

A list of metadata items that you can associate with your connection resource. A tag is a key-value pair.

createConnection_connectionName :: Lens' CreateConnection Text

A name for the new connection. It must be unique across all App Runner connections for the Amazon Web Services account in the Amazon Web Services Region.

createConnection_providerType :: Lens' CreateConnection ProviderType

The source repository provider.

createConnectionResponse_httpStatus :: Lens' CreateConnectionResponse Int

The response's http status code.

createConnectionResponse_connection :: Lens' CreateConnectionResponse Connection

A description of the App Runner connection that's created by this request.

DescribeCustomDomains

describeCustomDomains_nextToken :: Lens' DescribeCustomDomains (Maybe Text)

A token from a previous result page. It's used for a paginated request. The request retrieves the next result page. All other parameter values must be identical to the ones that are specified in the initial request.

If you don't specify NextToken, the request retrieves the first result page.

describeCustomDomains_maxResults :: Lens' DescribeCustomDomains (Maybe Natural)

The maximum number of results that each response (result page) can include. It's used for a paginated request.

If you don't specify MaxResults, the request retrieves all available results in a single response.

describeCustomDomains_serviceArn :: Lens' DescribeCustomDomains Text

The Amazon Resource Name (ARN) of the App Runner service that you want associated custom domain names to be described for.

describeCustomDomainsResponse_nextToken :: Lens' DescribeCustomDomainsResponse (Maybe Text)

The token that you can pass in a subsequent request to get the next result page. It's returned in a paginated request.

describeCustomDomainsResponse_httpStatus :: Lens' DescribeCustomDomainsResponse Int

The response's http status code.

describeCustomDomainsResponse_dNSTarget :: Lens' DescribeCustomDomainsResponse Text

The App Runner subdomain of the App Runner service. The associated custom domain names are mapped to this target name.

describeCustomDomainsResponse_serviceArn :: Lens' DescribeCustomDomainsResponse Text

The Amazon Resource Name (ARN) of the App Runner service whose associated custom domain names you want to describe.

describeCustomDomainsResponse_customDomains :: Lens' DescribeCustomDomainsResponse [CustomDomain]

A list of descriptions of custom domain names that are associated with the service. In a paginated request, the request returns up to MaxResults records per call.

DescribeService

describeService_serviceArn :: Lens' DescribeService Text

The Amazon Resource Name (ARN) of the App Runner service that you want a description for.

describeServiceResponse_httpStatus :: Lens' DescribeServiceResponse Int

The response's http status code.

describeServiceResponse_service :: Lens' DescribeServiceResponse Service

A full description of the App Runner service that you specified in this request.

DeleteAutoScalingConfiguration

deleteAutoScalingConfiguration_autoScalingConfigurationArn :: Lens' DeleteAutoScalingConfiguration Text

The Amazon Resource Name (ARN) of the App Runner auto scaling configuration that you want to delete.

The ARN can be a full auto scaling configuration ARN, or a partial ARN ending with either .../name or .../name/revision . If a revision isn't specified, the latest active revision is deleted.

deleteAutoScalingConfigurationResponse_httpStatus :: Lens' DeleteAutoScalingConfigurationResponse Int

The response's http status code.

deleteAutoScalingConfigurationResponse_autoScalingConfiguration :: Lens' DeleteAutoScalingConfigurationResponse AutoScalingConfiguration

A description of the App Runner auto scaling configuration that this request just deleted.

ListAutoScalingConfigurations

listAutoScalingConfigurations_autoScalingConfigurationName :: Lens' ListAutoScalingConfigurations (Maybe Text)

The name of the App Runner auto scaling configuration that you want to list. If specified, App Runner lists revisions that share this name. If not specified, App Runner returns revisions of all configurations.

listAutoScalingConfigurations_nextToken :: Lens' ListAutoScalingConfigurations (Maybe Text)

A token from a previous result page. It's used for a paginated request. The request retrieves the next result page. All other parameter values must be identical to the ones that are specified in the initial request.

If you don't specify NextToken, the request retrieves the first result page.

listAutoScalingConfigurations_latestOnly :: Lens' ListAutoScalingConfigurations (Maybe Bool)

Set to true to list only the latest revision for each requested configuration name.

Keep as false to list all revisions for each requested configuration name.

Default: false

listAutoScalingConfigurations_maxResults :: Lens' ListAutoScalingConfigurations (Maybe Natural)

The maximum number of results to include in each response (result page). It's used for a paginated request.

If you don't specify MaxResults, the request retrieves all available results in a single response.

listAutoScalingConfigurationsResponse_nextToken :: Lens' ListAutoScalingConfigurationsResponse (Maybe Text)

The token that you can pass in a subsequent request to get the next result page. It's returned in a paginated request.

listAutoScalingConfigurationsResponse_httpStatus :: Lens' ListAutoScalingConfigurationsResponse Int

The response's http status code.

listAutoScalingConfigurationsResponse_autoScalingConfigurationSummaryList :: Lens' ListAutoScalingConfigurationsResponse [AutoScalingConfigurationSummary]

A list of summary information records for auto scaling configurations. In a paginated request, the request returns up to MaxResults records for each call.

DisassociateCustomDomain

disassociateCustomDomain_serviceArn :: Lens' DisassociateCustomDomain Text

The Amazon Resource Name (ARN) of the App Runner service that you want to disassociate a custom domain name from.

disassociateCustomDomain_domainName :: Lens' DisassociateCustomDomain Text

The domain name that you want to disassociate from the App Runner service.

disassociateCustomDomainResponse_httpStatus :: Lens' DisassociateCustomDomainResponse Int

The response's http status code.

disassociateCustomDomainResponse_dNSTarget :: Lens' DisassociateCustomDomainResponse Text

The App Runner subdomain of the App Runner service. The disassociated custom domain name was mapped to this target name.

disassociateCustomDomainResponse_serviceArn :: Lens' DisassociateCustomDomainResponse Text

The Amazon Resource Name (ARN) of the App Runner service that a custom domain name is disassociated from.

disassociateCustomDomainResponse_customDomain :: Lens' DisassociateCustomDomainResponse CustomDomain

A description of the domain name that's being disassociated.

PauseService

pauseService_serviceArn :: Lens' PauseService Text

The Amazon Resource Name (ARN) of the App Runner service that you want to pause.

pauseServiceResponse_operationId :: Lens' PauseServiceResponse (Maybe Text)

The unique ID of the asynchronous operation that this request started. You can use it combined with the ListOperations call to track the operation's progress.

pauseServiceResponse_httpStatus :: Lens' PauseServiceResponse Int

The response's http status code.

pauseServiceResponse_service :: Lens' PauseServiceResponse Service

A description of the App Runner service that this request just paused.

TagResource

tagResource_resourceArn :: Lens' TagResource Text

The Amazon Resource Name (ARN) of the resource that you want to update tags for.

It must be the ARN of an App Runner resource.

tagResource_tags :: Lens' TagResource [Tag]

A list of tag key-value pairs to add or update. If a key is new to the resource, the tag is added with the provided value. If a key is already associated with the resource, the value of the tag is updated.

tagResourceResponse_httpStatus :: Lens' TagResourceResponse Int

The response's http status code.

UntagResource

untagResource_resourceArn :: Lens' UntagResource Text

The Amazon Resource Name (ARN) of the resource that you want to remove tags from.

It must be the ARN of an App Runner resource.

untagResource_tagKeys :: Lens' UntagResource [Text]

A list of tag keys that you want to remove.

untagResourceResponse_httpStatus :: Lens' UntagResourceResponse Int

The response's http status code.

CreateAutoScalingConfiguration

createAutoScalingConfiguration_maxSize :: Lens' CreateAutoScalingConfiguration (Maybe Natural)

The maximum number of instances that your service scales up to. At most MaxSize instances actively serve traffic for your service.

Default: 25

createAutoScalingConfiguration_minSize :: Lens' CreateAutoScalingConfiguration (Maybe Natural)

The minimum number of instances that App Runner provisions for your service. The service always has at least MinSize provisioned instances. Some of them actively serve traffic. The rest of them (provisioned and inactive instances) are a cost-effective compute capacity reserve and are ready to be quickly activated. You pay for memory usage of all the provisioned instances. You pay for CPU usage of only the active subset.

App Runner temporarily doubles the number of provisioned instances during deployments, to maintain the same capacity for both old and new code.

Default: 1

createAutoScalingConfiguration_tags :: Lens' CreateAutoScalingConfiguration (Maybe [Tag])

A list of metadata items that you can associate with your auto scaling configuration resource. A tag is a key-value pair.

createAutoScalingConfiguration_maxConcurrency :: Lens' CreateAutoScalingConfiguration (Maybe Natural)

The maximum number of concurrent requests that you want an instance to process. If the number of concurrent requests exceeds this limit, App Runner scales up your service.

Default: 100

createAutoScalingConfiguration_autoScalingConfigurationName :: Lens' CreateAutoScalingConfiguration Text

A name for the auto scaling configuration. When you use it for the first time in an Amazon Web Services Region, App Runner creates revision number 1 of this name. When you use the same name in subsequent calls, App Runner creates incremental revisions of the configuration.

createAutoScalingConfigurationResponse_httpStatus :: Lens' CreateAutoScalingConfigurationResponse Int

The response's http status code.

createAutoScalingConfigurationResponse_autoScalingConfiguration :: Lens' CreateAutoScalingConfigurationResponse AutoScalingConfiguration

A description of the App Runner auto scaling configuration that's created by this request.

StartDeployment

startDeployment_serviceArn :: Lens' StartDeployment Text

The Amazon Resource Name (ARN) of the App Runner service that you want to manually deploy to.

startDeploymentResponse_httpStatus :: Lens' StartDeploymentResponse Int

The response's http status code.

startDeploymentResponse_operationId :: Lens' StartDeploymentResponse Text

The unique ID of the asynchronous operation that this request started. You can use it combined with the ListOperations call to track the operation's progress.

CreateService

createService_autoScalingConfigurationArn :: Lens' CreateService (Maybe Text)

The Amazon Resource Name (ARN) of an App Runner automatic scaling configuration resource that you want to associate with your service. If not provided, App Runner associates the latest revision of a default auto scaling configuration.

createService_encryptionConfiguration :: Lens' CreateService (Maybe EncryptionConfiguration)

An optional custom encryption key that App Runner uses to encrypt the copy of your source repository that it maintains and your service logs. By default, App Runner uses an Amazon Web Services managed CMK.

createService_healthCheckConfiguration :: Lens' CreateService (Maybe HealthCheckConfiguration)

The settings for the health check that App Runner performs to monitor the health of your service.

createService_tags :: Lens' CreateService (Maybe [Tag])

An optional list of metadata items that you can associate with your service resource. A tag is a key-value pair.

createService_instanceConfiguration :: Lens' CreateService (Maybe InstanceConfiguration)

The runtime configuration of instances (scaling units) of the App Runner service.

createService_serviceName :: Lens' CreateService Text

A name for the new service. It must be unique across all the running App Runner services in your Amazon Web Services account in the Amazon Web Services Region.

createService_sourceConfiguration :: Lens' CreateService SourceConfiguration

The source to deploy to the App Runner service. It can be a code or an image repository.

createServiceResponse_httpStatus :: Lens' CreateServiceResponse Int

The response's http status code.

createServiceResponse_service :: Lens' CreateServiceResponse Service

A description of the App Runner service that's created by this request.

createServiceResponse_operationId :: Lens' CreateServiceResponse Text

The unique ID of the asynchronous operation that this request started. You can use it combined with the ListOperations call to track the operation's progress.

Types

AuthenticationConfiguration

authenticationConfiguration_accessRoleArn :: Lens' AuthenticationConfiguration (Maybe Text)

The Amazon Resource Name (ARN) of the IAM role that grants the App Runner service access to a source repository. It's required for ECR image repositories (but not for ECR Public repositories).

authenticationConfiguration_connectionArn :: Lens' AuthenticationConfiguration (Maybe Text)

The Amazon Resource Name (ARN) of the App Runner connection that enables the App Runner service to connect to a source repository. It's required for GitHub code repositories.

AutoScalingConfiguration

autoScalingConfiguration_status :: Lens' AutoScalingConfiguration (Maybe AutoScalingConfigurationStatus)

The current state of the auto scaling configuration. If the status of a configuration revision is INACTIVE, it was deleted and can't be used. Inactive configuration revisions are permanently removed some time after they are deleted.

autoScalingConfiguration_autoScalingConfigurationName :: Lens' AutoScalingConfiguration (Maybe Text)

The customer-provided auto scaling configuration name. It can be used in multiple revisions of a configuration.

autoScalingConfiguration_createdAt :: Lens' AutoScalingConfiguration (Maybe UTCTime)

The time when the auto scaling configuration was created. It's in Unix time stamp format.

autoScalingConfiguration_maxSize :: Lens' AutoScalingConfiguration (Maybe Int)

The maximum number of instances that a service scales up to. At most MaxSize instances actively serve traffic for your service.

autoScalingConfiguration_autoScalingConfigurationRevision :: Lens' AutoScalingConfiguration (Maybe Int)

The revision of this auto scaling configuration. It's unique among all the active configurations ("Status": "ACTIVE") that share the same AutoScalingConfigurationName.

autoScalingConfiguration_autoScalingConfigurationArn :: Lens' AutoScalingConfiguration (Maybe Text)

The Amazon Resource Name (ARN) of this auto scaling configuration.

autoScalingConfiguration_minSize :: Lens' AutoScalingConfiguration (Maybe Int)

The minimum number of instances that App Runner provisions for a service. The service always has at least MinSize provisioned instances. Some of them actively serve traffic. The rest of them (provisioned and inactive instances) are a cost-effective compute capacity reserve and are ready to be quickly activated. You pay for memory usage of all the provisioned instances. You pay for CPU usage of only the active subset.

App Runner temporarily doubles the number of provisioned instances during deployments, to maintain the same capacity for both old and new code.

autoScalingConfiguration_deletedAt :: Lens' AutoScalingConfiguration (Maybe UTCTime)

The time when the auto scaling configuration was deleted. It's in Unix time stamp format.

autoScalingConfiguration_latest :: Lens' AutoScalingConfiguration (Maybe Bool)

It's set to true for the configuration with the highest Revision among all configurations that share the same Name. It's set to false otherwise.

autoScalingConfiguration_maxConcurrency :: Lens' AutoScalingConfiguration (Maybe Int)

The maximum number of concurrent requests that an instance processes. If the number of concurrent requests exceeds this limit, App Runner scales the service up.

AutoScalingConfigurationSummary

autoScalingConfigurationSummary_autoScalingConfigurationName :: Lens' AutoScalingConfigurationSummary (Maybe Text)

The customer-provided auto scaling configuration name. It can be used in multiple revisions of a configuration.

autoScalingConfigurationSummary_autoScalingConfigurationRevision :: Lens' AutoScalingConfigurationSummary (Maybe Int)

The revision of this auto scaling configuration. It's unique among all the active configurations ("Status": "ACTIVE") with the same AutoScalingConfigurationName.

autoScalingConfigurationSummary_autoScalingConfigurationArn :: Lens' AutoScalingConfigurationSummary (Maybe Text)

The Amazon Resource Name (ARN) of this auto scaling configuration.

CertificateValidationRecord

certificateValidationRecord_status :: Lens' CertificateValidationRecord (Maybe CertificateValidationRecordStatus)

The current state of the certificate CNAME record validation. It should change to SUCCESS after App Runner completes validation with your DNS.

certificateValidationRecord_value :: Lens' CertificateValidationRecord (Maybe Text)

The certificate CNAME record value.

certificateValidationRecord_name :: Lens' CertificateValidationRecord (Maybe Text)

The certificate CNAME record name.

certificateValidationRecord_type :: Lens' CertificateValidationRecord (Maybe Text)

The record type, always CNAME.

CodeConfiguration

codeConfiguration_codeConfigurationValues :: Lens' CodeConfiguration (Maybe CodeConfigurationValues)

The basic configuration for building and running the App Runner service. Use it to quickly launch an App Runner service without providing a apprunner.yaml file in the source code repository (or ignoring the file if it exists).

codeConfiguration_configurationSource :: Lens' CodeConfiguration ConfigurationSource

The source of the App Runner configuration. Values are interpreted as follows:

REPOSITORY – App Runner reads configuration values from the apprunner.yaml file in the source code repository and ignores CodeConfigurationValues.

API – App Runner uses configuration values provided in CodeConfigurationValues and ignores the apprunner.yaml file in the source code repository.

CodeConfigurationValues

codeConfigurationValues_startCommand :: Lens' CodeConfigurationValues (Maybe Text)

The command App Runner runs to start your application.

codeConfigurationValues_runtimeEnvironmentVariables :: Lens' CodeConfigurationValues (Maybe (HashMap Text Text))

The environment variables that are available to your running App Runner service. An array of key-value pairs. Keys with a prefix of AWSAPPRUNNER are reserved for system use and aren't valid.

codeConfigurationValues_buildCommand :: Lens' CodeConfigurationValues (Maybe Text)

The command App Runner runs to build your application.

codeConfigurationValues_port :: Lens' CodeConfigurationValues (Maybe Text)

The port that your application listens to in the container.

Default: 8080

codeConfigurationValues_runtime :: Lens' CodeConfigurationValues Runtime

A runtime environment type for building and running an App Runner service. It represents a programming language runtime.

CodeRepository

codeRepository_codeConfiguration :: Lens' CodeRepository (Maybe CodeConfiguration)

Configuration for building and running the service from a source code repository.

codeRepository_repositoryUrl :: Lens' CodeRepository Text

The location of the repository that contains the source code.

codeRepository_sourceCodeVersion :: Lens' CodeRepository SourceCodeVersion

The version that should be used within the source code repository.

Connection

connection_status :: Lens' Connection (Maybe ConnectionStatus)

The current state of the App Runner connection. When the state is AVAILABLE, you can use the connection to create an App Runner service.

connection_createdAt :: Lens' Connection (Maybe UTCTime)

The App Runner connection creation time, expressed as a Unix time stamp.

connection_providerType :: Lens' Connection (Maybe ProviderType)

The source repository provider.

connection_connectionName :: Lens' Connection (Maybe Text)

The customer-provided connection name.

connection_connectionArn :: Lens' Connection (Maybe Text)

The Amazon Resource Name (ARN) of this connection.

ConnectionSummary

connectionSummary_status :: Lens' ConnectionSummary (Maybe ConnectionStatus)

The current state of the App Runner connection. When the state is AVAILABLE, you can use the connection to create an App Runner service.

connectionSummary_createdAt :: Lens' ConnectionSummary (Maybe UTCTime)

The App Runner connection creation time, expressed as a Unix time stamp.

connectionSummary_providerType :: Lens' ConnectionSummary (Maybe ProviderType)

The source repository provider.

connectionSummary_connectionName :: Lens' ConnectionSummary (Maybe Text)

The customer-provided connection name.

connectionSummary_connectionArn :: Lens' ConnectionSummary (Maybe Text)

The Amazon Resource Name (ARN) of this connection.

CustomDomain

customDomain_certificateValidationRecords :: Lens' CustomDomain (Maybe [CertificateValidationRecord])

A list of certificate CNAME records that's used for this domain name.

customDomain_domainName :: Lens' CustomDomain Text

An associated custom domain endpoint. It can be a root domain (for example, example.com), a subdomain (for example, login.example.com or admin.login.example.com), or a wildcard (for example, *.example.com).

customDomain_enableWWWSubdomain :: Lens' CustomDomain Bool

When true, the subdomain www.DomainName is associated with the App Runner service in addition to the base domain.

customDomain_status :: Lens' CustomDomain CustomDomainAssociationStatus

The current state of the domain name association.

EncryptionConfiguration

encryptionConfiguration_kmsKey :: Lens' EncryptionConfiguration Text

The ARN of the KMS key that's used for encryption.

HealthCheckConfiguration

healthCheckConfiguration_healthyThreshold :: Lens' HealthCheckConfiguration (Maybe Natural)

The number of consecutive checks that must succeed before App Runner decides that the service is healthy.

Default: 1

healthCheckConfiguration_path :: Lens' HealthCheckConfiguration (Maybe Text)

The URL that health check requests are sent to.

Path is only applicable when you set Protocol to HTTP.

Default: "/"

healthCheckConfiguration_protocol :: Lens' HealthCheckConfiguration (Maybe HealthCheckProtocol)

The IP protocol that App Runner uses to perform health checks for your service.

If you set Protocol to HTTP, App Runner sends health check requests to the HTTP path specified by Path.

Default: TCP

healthCheckConfiguration_interval :: Lens' HealthCheckConfiguration (Maybe Natural)

The time interval, in seconds, between health checks.

Default: 5

healthCheckConfiguration_timeout :: Lens' HealthCheckConfiguration (Maybe Natural)

The time, in seconds, to wait for a health check response before deciding it failed.

Default: 2

healthCheckConfiguration_unhealthyThreshold :: Lens' HealthCheckConfiguration (Maybe Natural)

The number of consecutive checks that must fail before App Runner decides that the service is unhealthy.

Default: 5

ImageConfiguration

imageConfiguration_startCommand :: Lens' ImageConfiguration (Maybe Text)

An optional command that App Runner runs to start the application in the source image. If specified, this command overrides the Docker image’s default start command.

imageConfiguration_runtimeEnvironmentVariables :: Lens' ImageConfiguration (Maybe (HashMap Text Text))

Environment variables that are available to your running App Runner service. An array of key-value pairs. Keys with a prefix of AWSAPPRUNNER are reserved for system use and aren't valid.

imageConfiguration_port :: Lens' ImageConfiguration (Maybe Text)

The port that your application listens to in the container.

Default: 8080

ImageRepository

imageRepository_imageConfiguration :: Lens' ImageRepository (Maybe ImageConfiguration)

Configuration for running the identified image.

imageRepository_imageIdentifier :: Lens' ImageRepository Text

The identifier of an image.

For an image in Amazon Elastic Container Registry (Amazon ECR), this is an image name. For the image name format, see Pulling an image in the Amazon ECR User Guide.

imageRepository_imageRepositoryType :: Lens' ImageRepository ImageRepositoryType

The type of the image repository. This reflects the repository provider and whether the repository is private or public.

InstanceConfiguration

instanceConfiguration_memory :: Lens' InstanceConfiguration (Maybe Text)

The amount of memory, in MB or GB, reserved for each instance of your App Runner service.

Default: 2 GB

instanceConfiguration_instanceRoleArn :: Lens' InstanceConfiguration (Maybe Text)

The Amazon Resource Name (ARN) of an IAM role that provides permissions to your App Runner service. These are permissions that your code needs when it calls any Amazon Web Services APIs.

instanceConfiguration_cpu :: Lens' InstanceConfiguration (Maybe Text)

The number of CPU units reserved for each instance of your App Runner service.

Default: 1 vCPU

OperationSummary

operationSummary_status :: Lens' OperationSummary (Maybe OperationStatus)

The current state of the operation.

operationSummary_targetArn :: Lens' OperationSummary (Maybe Text)

The Amazon Resource Name (ARN) of the resource that the operation acted on (for example, an App Runner service).

operationSummary_endedAt :: Lens' OperationSummary (Maybe UTCTime)

The time when the operation ended. It's in the Unix time stamp format.

operationSummary_startedAt :: Lens' OperationSummary (Maybe UTCTime)

The time when the operation started. It's in the Unix time stamp format.

operationSummary_id :: Lens' OperationSummary (Maybe Text)

A unique ID of this operation. It's unique in the scope of the App Runner service.

operationSummary_type :: Lens' OperationSummary (Maybe OperationType)

The type of operation. It indicates a specific action that occured.

operationSummary_updatedAt :: Lens' OperationSummary (Maybe UTCTime)

The time when the operation was last updated. It's in the Unix time stamp format.

Service

service_encryptionConfiguration :: Lens' Service (Maybe EncryptionConfiguration)

The encryption key that App Runner uses to encrypt the service logs and the copy of the source repository that App Runner maintains for the service. It can be either a customer-provided encryption key or an Amazon Web Services managed CMK.

service_healthCheckConfiguration :: Lens' Service (Maybe HealthCheckConfiguration)

The settings for the health check that App Runner performs to monitor the health of this service.

service_deletedAt :: Lens' Service (Maybe UTCTime)

The time when the App Runner service was deleted. It's in the Unix time stamp format.

service_serviceName :: Lens' Service Text

The customer-provided service name.

service_serviceId :: Lens' Service Text

An ID that App Runner generated for this service. It's unique within the Amazon Web Services Region.

service_serviceArn :: Lens' Service Text

The Amazon Resource Name (ARN) of this service.

service_serviceUrl :: Lens' Service Text

A subdomain URL that App Runner generated for this service. You can use this URL to access your service web application.

service_createdAt :: Lens' Service UTCTime

The time when the App Runner service was created. It's in the Unix time stamp format.

service_updatedAt :: Lens' Service UTCTime

The time when the App Runner service was last updated at. It's in the Unix time stamp format.

service_status :: Lens' Service ServiceStatus

The current state of the App Runner service. These particular values mean the following.

CREATE_FAILED – The service failed to create. To troubleshoot this failure, read the failure events and logs, change any parameters that need to be fixed, and retry the call to create the service.

The failed service isn't usable, and still counts towards your service quota. When you're done analyzing the failure, delete the service.

DELETE_FAILED – The service failed to delete and can't be successfully recovered. Retry the service deletion call to ensure that all related resources are removed.

service_sourceConfiguration :: Lens' Service SourceConfiguration

The source deployed to the App Runner service. It can be a code or an image repository.

service_instanceConfiguration :: Lens' Service InstanceConfiguration

The runtime configuration of instances (scaling units) of this service.

service_autoScalingConfigurationSummary :: Lens' Service AutoScalingConfigurationSummary

Summary information for the App Runner automatic scaling configuration resource that's associated with this service.

ServiceSummary

serviceSummary_status :: Lens' ServiceSummary (Maybe ServiceStatus)

The current state of the App Runner service. These particular values mean the following.

CREATE_FAILED – The service failed to create. Read the failure events and logs, change any parameters that need to be fixed, and retry the call to create the service.

The failed service isn't usable, and still counts towards your service quota. When you're done analyzing the failure, delete the service.

DELETE_FAILED – The service failed to delete and can't be successfully recovered. Retry the service deletion call to ensure that all related resources are removed.

serviceSummary_createdAt :: Lens' ServiceSummary (Maybe UTCTime)

The time when the App Runner service was created. It's in the Unix time stamp format.

serviceSummary_serviceUrl :: Lens' ServiceSummary (Maybe Text)

A subdomain URL that App Runner generated for this service. You can use this URL to access your service web application.

serviceSummary_serviceName :: Lens' ServiceSummary (Maybe Text)

The customer-provided service name.

serviceSummary_updatedAt :: Lens' ServiceSummary (Maybe UTCTime)

The time when the App Runner service was last updated. It's in theUnix time stamp format.

serviceSummary_serviceArn :: Lens' ServiceSummary (Maybe Text)

The Amazon Resource Name (ARN) of this service.

serviceSummary_serviceId :: Lens' ServiceSummary (Maybe Text)

An ID that App Runner generated for this service. It's unique within the Amazon Web Services Region.

SourceCodeVersion

sourceCodeVersion_type :: Lens' SourceCodeVersion SourceCodeVersionType

The type of version identifier.

For a git-based repository, branches represent versions.

sourceCodeVersion_value :: Lens' SourceCodeVersion Text

A source code version.

For a git-based repository, a branch name maps to a specific version. App Runner uses the most recent commit to the branch.

SourceConfiguration

sourceConfiguration_imageRepository :: Lens' SourceConfiguration (Maybe ImageRepository)

The description of a source image repository.

You must provide either this member or CodeRepository (but not both).

sourceConfiguration_codeRepository :: Lens' SourceConfiguration (Maybe CodeRepository)

The description of a source code repository.

You must provide either this member or ImageRepository (but not both).

sourceConfiguration_autoDeploymentsEnabled :: Lens' SourceConfiguration (Maybe Bool)

If true, continuous integration from the source repository is enabled for the App Runner service. Each repository change (including any source code commit or new image version) starts a deployment.

Default: App Runner sets to false for a source image that uses an ECR Public repository or an ECR repository that's in an Amazon Web Services account other than the one that the service is in. App Runner sets to true in all other cases (which currently include a source code repository or a source image using a same-account ECR repository).

sourceConfiguration_authenticationConfiguration :: Lens' SourceConfiguration (Maybe AuthenticationConfiguration)

Describes the resources that are needed to authenticate access to some source repositories.

Tag

tag_value :: Lens' Tag (Maybe Text)

The value of the tag.

tag_key :: Lens' Tag (Maybe Text)

The key of the tag.