Amazonka.AppRunner.ListAutoScalingConfigurations

Description

Returns a list of App Runner automatic scaling configurations in your Amazon Web Services account. You can query the revisions for a specific configuration name or the revisions for all configurations in your account. You can optionally query only the latest revision of each requested name.

Creating a Request

data ListAutoScalingConfigurations

See: newListAutoScalingConfigurations smart constructor.

Constructors

ListAutoScalingConfigurations'

Fields

autoScalingConfigurationName :: Maybe Text

The name of the App Runner auto scaling configuration that you want to list. If specified, App Runner lists revisions that share this name. If not specified, App Runner returns revisions of all configurations.

nextToken :: Maybe Text

A token from a previous result page. It's used for a paginated request. The request retrieves the next result page. All other parameter values must be identical to the ones that are specified in the initial request.

If you don't specify NextToken, the request retrieves the first result page.

latestOnly :: Maybe Bool

Set to true to list only the latest revision for each requested configuration name.

Keep as false to list all revisions for each requested configuration name.

Default: false

maxResults :: Maybe Natural

The maximum number of results to include in each response (result page). It's used for a paginated request.

If you don't specify MaxResults, the request retrieves all available results in a single response.

newListAutoScalingConfigurations :: ListAutoScalingConfigurations

Create a value of ListAutoScalingConfigurations with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:autoScalingConfigurationName:ListAutoScalingConfigurations', listAutoScalingConfigurations_autoScalingConfigurationName - The name of the App Runner auto scaling configuration that you want to list. If specified, App Runner lists revisions that share this name. If not specified, App Runner returns revisions of all configurations.

$sel:nextToken:ListAutoScalingConfigurations', listAutoScalingConfigurations_nextToken - A token from a previous result page. It's used for a paginated request. The request retrieves the next result page. All other parameter values must be identical to the ones that are specified in the initial request.

If you don't specify NextToken, the request retrieves the first result page.

$sel:latestOnly:ListAutoScalingConfigurations', listAutoScalingConfigurations_latestOnly - Set to true to list only the latest revision for each requested configuration name.

Keep as false to list all revisions for each requested configuration name.

Default: false

$sel:maxResults:ListAutoScalingConfigurations', listAutoScalingConfigurations_maxResults - The maximum number of results to include in each response (result page). It's used for a paginated request.

If you don't specify MaxResults, the request retrieves all available results in a single response.

Request Lenses

listAutoScalingConfigurations_autoScalingConfigurationName :: Lens' ListAutoScalingConfigurations (Maybe Text)

The name of the App Runner auto scaling configuration that you want to list. If specified, App Runner lists revisions that share this name. If not specified, App Runner returns revisions of all configurations.

listAutoScalingConfigurations_nextToken :: Lens' ListAutoScalingConfigurations (Maybe Text)

A token from a previous result page. It's used for a paginated request. The request retrieves the next result page. All other parameter values must be identical to the ones that are specified in the initial request.

If you don't specify NextToken, the request retrieves the first result page.

listAutoScalingConfigurations_latestOnly :: Lens' ListAutoScalingConfigurations (Maybe Bool)

Set to true to list only the latest revision for each requested configuration name.

Keep as false to list all revisions for each requested configuration name.

Default: false

listAutoScalingConfigurations_maxResults :: Lens' ListAutoScalingConfigurations (Maybe Natural)

The maximum number of results to include in each response (result page). It's used for a paginated request.

If you don't specify MaxResults, the request retrieves all available results in a single response.

Destructuring the Response

data ListAutoScalingConfigurationsResponse

See: newListAutoScalingConfigurationsResponse smart constructor.

Constructors

ListAutoScalingConfigurationsResponse'

Fields

nextToken :: Maybe Text

The token that you can pass in a subsequent request to get the next result page. It's returned in a paginated request.

httpStatus :: Int

The response's http status code.

autoScalingConfigurationSummaryList :: [AutoScalingConfigurationSummary]

A list of summary information records for auto scaling configurations. In a paginated request, the request returns up to MaxResults records for each call.

newListAutoScalingConfigurationsResponse

Arguments

:: Int
$sel:httpStatus:ListAutoScalingConfigurationsResponse'

-> ListAutoScalingConfigurationsResponse

Create a value of ListAutoScalingConfigurationsResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:nextToken:ListAutoScalingConfigurations', listAutoScalingConfigurationsResponse_nextToken - The token that you can pass in a subsequent request to get the next result page. It's returned in a paginated request.

$sel:httpStatus:ListAutoScalingConfigurationsResponse', listAutoScalingConfigurationsResponse_httpStatus - The response's http status code.

$sel:autoScalingConfigurationSummaryList:ListAutoScalingConfigurationsResponse', listAutoScalingConfigurationsResponse_autoScalingConfigurationSummaryList - A list of summary information records for auto scaling configurations. In a paginated request, the request returns up to MaxResults records for each call.

Response Lenses

listAutoScalingConfigurationsResponse_nextToken :: Lens' ListAutoScalingConfigurationsResponse (Maybe Text)

The token that you can pass in a subsequent request to get the next result page. It's returned in a paginated request.

listAutoScalingConfigurationsResponse_httpStatus :: Lens' ListAutoScalingConfigurationsResponse Int

The response's http status code.

listAutoScalingConfigurationsResponse_autoScalingConfigurationSummaryList :: Lens' ListAutoScalingConfigurationsResponse [AutoScalingConfigurationSummary]

A list of summary information records for auto scaling configurations. In a paginated request, the request returns up to MaxResults records for each call.