Amazonka.AppRunner.PauseService

Description

Pause an active App Runner service. App Runner reduces compute capacity for the service to zero and loses state (for example, ephemeral storage is removed).

This is an asynchronous operation. On a successful call, you can use the returned OperationId and the ListOperations call to track the operation's progress.

Creating a Request

data PauseService

See: newPauseService smart constructor.

Constructors

PauseService'

Fields

serviceArn :: Text

The Amazon Resource Name (ARN) of the App Runner service that you want to pause.

newPauseService

Arguments

:: Text
$sel:serviceArn:PauseService'

-> PauseService

Create a value of PauseService with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:serviceArn:PauseService', pauseService_serviceArn - The Amazon Resource Name (ARN) of the App Runner service that you want to pause.

Request Lenses

pauseService_serviceArn :: Lens' PauseService Text

The Amazon Resource Name (ARN) of the App Runner service that you want to pause.

Destructuring the Response

data PauseServiceResponse

See: newPauseServiceResponse smart constructor.

Constructors

PauseServiceResponse'

Fields

operationId :: Maybe Text

The unique ID of the asynchronous operation that this request started. You can use it combined with the ListOperations call to track the operation's progress.

httpStatus :: Int

The response's http status code.

service :: Service

A description of the App Runner service that this request just paused.

newPauseServiceResponse

Arguments

:: Int
$sel:httpStatus:PauseServiceResponse'

-> Service
$sel:service:PauseServiceResponse'

-> PauseServiceResponse

Create a value of PauseServiceResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:operationId:PauseServiceResponse', pauseServiceResponse_operationId - The unique ID of the asynchronous operation that this request started. You can use it combined with the ListOperations call to track the operation's progress.

$sel:httpStatus:PauseServiceResponse', pauseServiceResponse_httpStatus - The response's http status code.

$sel:service:PauseServiceResponse', pauseServiceResponse_service - A description of the App Runner service that this request just paused.

Response Lenses

pauseServiceResponse_operationId :: Lens' PauseServiceResponse (Maybe Text)

The unique ID of the asynchronous operation that this request started. You can use it combined with the ListOperations call to track the operation's progress.

pauseServiceResponse_httpStatus :: Lens' PauseServiceResponse Int

The response's http status code.

pauseServiceResponse_service :: Lens' PauseServiceResponse Service

A description of the App Runner service that this request just paused.