Amazonka.AppRunner.Types.AutoScalingConfiguration

Description

Documentation

data AutoScalingConfiguration

Describes an App Runner automatic scaling configuration resource. Multiple revisions of a configuration have the same AutoScalingConfigurationName and different AutoScalingConfigurationRevision values.

A higher MinSize increases the spread of your App Runner service over more Availability Zones in the Amazon Web Services Region. The tradeoff is a higher minimal cost.

A lower MaxSize controls your cost. The tradeoff is lower responsiveness during peak demand.

See: newAutoScalingConfiguration smart constructor.

Constructors

AutoScalingConfiguration'

Fields

status :: Maybe AutoScalingConfigurationStatus

The current state of the auto scaling configuration. If the status of a configuration revision is INACTIVE, it was deleted and can't be used. Inactive configuration revisions are permanently removed some time after they are deleted.

autoScalingConfigurationName :: Maybe Text

The customer-provided auto scaling configuration name. It can be used in multiple revisions of a configuration.

createdAt :: Maybe POSIX

The time when the auto scaling configuration was created. It's in Unix time stamp format.

maxSize :: Maybe Int

The maximum number of instances that a service scales up to. At most MaxSize instances actively serve traffic for your service.

autoScalingConfigurationRevision :: Maybe Int

The revision of this auto scaling configuration. It's unique among all the active configurations ("Status": "ACTIVE") that share the same AutoScalingConfigurationName.

autoScalingConfigurationArn :: Maybe Text

The Amazon Resource Name (ARN) of this auto scaling configuration.

minSize :: Maybe Int

The minimum number of instances that App Runner provisions for a service. The service always has at least MinSize provisioned instances. Some of them actively serve traffic. The rest of them (provisioned and inactive instances) are a cost-effective compute capacity reserve and are ready to be quickly activated. You pay for memory usage of all the provisioned instances. You pay for CPU usage of only the active subset.

App Runner temporarily doubles the number of provisioned instances during deployments, to maintain the same capacity for both old and new code.

deletedAt :: Maybe POSIX

The time when the auto scaling configuration was deleted. It's in Unix time stamp format.

latest :: Maybe Bool

It's set to true for the configuration with the highest Revision among all configurations that share the same Name. It's set to false otherwise.

maxConcurrency :: Maybe Int

The maximum number of concurrent requests that an instance processes. If the number of concurrent requests exceeds this limit, App Runner scales the service up.

newAutoScalingConfiguration :: AutoScalingConfiguration

Create a value of AutoScalingConfiguration with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:status:AutoScalingConfiguration', autoScalingConfiguration_status - The current state of the auto scaling configuration. If the status of a configuration revision is INACTIVE, it was deleted and can't be used. Inactive configuration revisions are permanently removed some time after they are deleted.

$sel:autoScalingConfigurationName:AutoScalingConfiguration', autoScalingConfiguration_autoScalingConfigurationName - The customer-provided auto scaling configuration name. It can be used in multiple revisions of a configuration.

$sel:createdAt:AutoScalingConfiguration', autoScalingConfiguration_createdAt - The time when the auto scaling configuration was created. It's in Unix time stamp format.

$sel:maxSize:AutoScalingConfiguration', autoScalingConfiguration_maxSize - The maximum number of instances that a service scales up to. At most MaxSize instances actively serve traffic for your service.

$sel:autoScalingConfigurationRevision:AutoScalingConfiguration', autoScalingConfiguration_autoScalingConfigurationRevision - The revision of this auto scaling configuration. It's unique among all the active configurations ("Status": "ACTIVE") that share the same AutoScalingConfigurationName.

$sel:autoScalingConfigurationArn:AutoScalingConfiguration', autoScalingConfiguration_autoScalingConfigurationArn - The Amazon Resource Name (ARN) of this auto scaling configuration.

$sel:minSize:AutoScalingConfiguration', autoScalingConfiguration_minSize - The minimum number of instances that App Runner provisions for a service. The service always has at least MinSize provisioned instances. Some of them actively serve traffic. The rest of them (provisioned and inactive instances) are a cost-effective compute capacity reserve and are ready to be quickly activated. You pay for memory usage of all the provisioned instances. You pay for CPU usage of only the active subset.

App Runner temporarily doubles the number of provisioned instances during deployments, to maintain the same capacity for both old and new code.

$sel:deletedAt:AutoScalingConfiguration', autoScalingConfiguration_deletedAt - The time when the auto scaling configuration was deleted. It's in Unix time stamp format.

$sel:latest:AutoScalingConfiguration', autoScalingConfiguration_latest - It's set to true for the configuration with the highest Revision among all configurations that share the same Name. It's set to false otherwise.

$sel:maxConcurrency:AutoScalingConfiguration', autoScalingConfiguration_maxConcurrency - The maximum number of concurrent requests that an instance processes. If the number of concurrent requests exceeds this limit, App Runner scales the service up.

autoScalingConfiguration_status :: Lens' AutoScalingConfiguration (Maybe AutoScalingConfigurationStatus)

The current state of the auto scaling configuration. If the status of a configuration revision is INACTIVE, it was deleted and can't be used. Inactive configuration revisions are permanently removed some time after they are deleted.

autoScalingConfiguration_autoScalingConfigurationName :: Lens' AutoScalingConfiguration (Maybe Text)

The customer-provided auto scaling configuration name. It can be used in multiple revisions of a configuration.

autoScalingConfiguration_createdAt :: Lens' AutoScalingConfiguration (Maybe UTCTime)

The time when the auto scaling configuration was created. It's in Unix time stamp format.

autoScalingConfiguration_maxSize :: Lens' AutoScalingConfiguration (Maybe Int)

The maximum number of instances that a service scales up to. At most MaxSize instances actively serve traffic for your service.

autoScalingConfiguration_autoScalingConfigurationRevision :: Lens' AutoScalingConfiguration (Maybe Int)

The revision of this auto scaling configuration. It's unique among all the active configurations ("Status": "ACTIVE") that share the same AutoScalingConfigurationName.

autoScalingConfiguration_autoScalingConfigurationArn :: Lens' AutoScalingConfiguration (Maybe Text)

The Amazon Resource Name (ARN) of this auto scaling configuration.

autoScalingConfiguration_minSize :: Lens' AutoScalingConfiguration (Maybe Int)

The minimum number of instances that App Runner provisions for a service. The service always has at least MinSize provisioned instances. Some of them actively serve traffic. The rest of them (provisioned and inactive instances) are a cost-effective compute capacity reserve and are ready to be quickly activated. You pay for memory usage of all the provisioned instances. You pay for CPU usage of only the active subset.

App Runner temporarily doubles the number of provisioned instances during deployments, to maintain the same capacity for both old and new code.

autoScalingConfiguration_deletedAt :: Lens' AutoScalingConfiguration (Maybe UTCTime)

The time when the auto scaling configuration was deleted. It's in Unix time stamp format.

autoScalingConfiguration_latest :: Lens' AutoScalingConfiguration (Maybe Bool)

It's set to true for the configuration with the highest Revision among all configurations that share the same Name. It's set to false otherwise.

autoScalingConfiguration_maxConcurrency :: Lens' AutoScalingConfiguration (Maybe Int)

The maximum number of concurrent requests that an instance processes. If the number of concurrent requests exceeds this limit, App Runner scales the service up.