Amazonka.AppRunner.Types.CodeConfiguration

Description

Documentation

data CodeConfiguration

Describes the configuration that App Runner uses to build and run an App Runner service from a source code repository.

See: newCodeConfiguration smart constructor.

Constructors

CodeConfiguration'

Fields

codeConfigurationValues :: Maybe CodeConfigurationValues

The basic configuration for building and running the App Runner service. Use it to quickly launch an App Runner service without providing a apprunner.yaml file in the source code repository (or ignoring the file if it exists).

configurationSource :: ConfigurationSource

The source of the App Runner configuration. Values are interpreted as follows:

REPOSITORY – App Runner reads configuration values from the apprunner.yaml file in the source code repository and ignores CodeConfigurationValues.

API – App Runner uses configuration values provided in CodeConfigurationValues and ignores the apprunner.yaml file in the source code repository.

newCodeConfiguration

Arguments

:: ConfigurationSource
$sel:configurationSource:CodeConfiguration'

-> CodeConfiguration

Create a value of CodeConfiguration with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:codeConfigurationValues:CodeConfiguration', codeConfiguration_codeConfigurationValues - The basic configuration for building and running the App Runner service. Use it to quickly launch an App Runner service without providing a apprunner.yaml file in the source code repository (or ignoring the file if it exists).

$sel:configurationSource:CodeConfiguration', codeConfiguration_configurationSource - The source of the App Runner configuration. Values are interpreted as follows:

REPOSITORY – App Runner reads configuration values from the apprunner.yaml file in the source code repository and ignores CodeConfigurationValues.

API – App Runner uses configuration values provided in CodeConfigurationValues and ignores the apprunner.yaml file in the source code repository.

codeConfiguration_codeConfigurationValues :: Lens' CodeConfiguration (Maybe CodeConfigurationValues)

The basic configuration for building and running the App Runner service. Use it to quickly launch an App Runner service without providing a apprunner.yaml file in the source code repository (or ignoring the file if it exists).

codeConfiguration_configurationSource :: Lens' CodeConfiguration ConfigurationSource

The source of the App Runner configuration. Values are interpreted as follows:

REPOSITORY – App Runner reads configuration values from the apprunner.yaml file in the source code repository and ignores CodeConfigurationValues.

API – App Runner uses configuration values provided in CodeConfigurationValues and ignores the apprunner.yaml file in the source code repository.