Amazonka.AppRunner.Types.CodeRepository

Description

Documentation

data CodeRepository

Describes a source code repository.

See: newCodeRepository smart constructor.

Constructors

CodeRepository'

Fields

codeConfiguration :: Maybe CodeConfiguration

Configuration for building and running the service from a source code repository.

repositoryUrl :: Text

The location of the repository that contains the source code.

sourceCodeVersion :: SourceCodeVersion

The version that should be used within the source code repository.

newCodeRepository

Arguments

:: Text
$sel:repositoryUrl:CodeRepository'

-> SourceCodeVersion
$sel:sourceCodeVersion:CodeRepository'

-> CodeRepository

Create a value of CodeRepository with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:codeConfiguration:CodeRepository', codeRepository_codeConfiguration - Configuration for building and running the service from a source code repository.

$sel:repositoryUrl:CodeRepository', codeRepository_repositoryUrl - The location of the repository that contains the source code.

$sel:sourceCodeVersion:CodeRepository', codeRepository_sourceCodeVersion - The version that should be used within the source code repository.

codeRepository_codeConfiguration :: Lens' CodeRepository (Maybe CodeConfiguration)

Configuration for building and running the service from a source code repository.

codeRepository_repositoryUrl :: Lens' CodeRepository Text

The location of the repository that contains the source code.

codeRepository_sourceCodeVersion :: Lens' CodeRepository SourceCodeVersion

The version that should be used within the source code repository.