Amazonka.AppRunner.Types.HealthCheckConfiguration

Description

Documentation

data HealthCheckConfiguration

Describes the settings for the health check that App Runner performs to monitor the health of a service.

See: newHealthCheckConfiguration smart constructor.

Constructors

HealthCheckConfiguration'

Fields

healthyThreshold :: Maybe Natural

The number of consecutive checks that must succeed before App Runner decides that the service is healthy.

Default: 1

path :: Maybe Text

The URL that health check requests are sent to.

Path is only applicable when you set Protocol to HTTP.

Default: "/"

protocol :: Maybe HealthCheckProtocol

The IP protocol that App Runner uses to perform health checks for your service.

If you set Protocol to HTTP, App Runner sends health check requests to the HTTP path specified by Path.

Default: TCP

interval :: Maybe Natural

The time interval, in seconds, between health checks.

Default: 5

timeout :: Maybe Natural

The time, in seconds, to wait for a health check response before deciding it failed.

Default: 2

unhealthyThreshold :: Maybe Natural

The number of consecutive checks that must fail before App Runner decides that the service is unhealthy.

Default: 5

newHealthCheckConfiguration :: HealthCheckConfiguration

Create a value of HealthCheckConfiguration with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:healthyThreshold:HealthCheckConfiguration', healthCheckConfiguration_healthyThreshold - The number of consecutive checks that must succeed before App Runner decides that the service is healthy.

Default: 1

$sel:path:HealthCheckConfiguration', healthCheckConfiguration_path - The URL that health check requests are sent to.

Path is only applicable when you set Protocol to HTTP.

Default: "/"

$sel:protocol:HealthCheckConfiguration', healthCheckConfiguration_protocol - The IP protocol that App Runner uses to perform health checks for your service.

If you set Protocol to HTTP, App Runner sends health check requests to the HTTP path specified by Path.

Default: TCP

$sel:interval:HealthCheckConfiguration', healthCheckConfiguration_interval - The time interval, in seconds, between health checks.

Default: 5

$sel:timeout:HealthCheckConfiguration', healthCheckConfiguration_timeout - The time, in seconds, to wait for a health check response before deciding it failed.

Default: 2

$sel:unhealthyThreshold:HealthCheckConfiguration', healthCheckConfiguration_unhealthyThreshold - The number of consecutive checks that must fail before App Runner decides that the service is unhealthy.

Default: 5

healthCheckConfiguration_healthyThreshold :: Lens' HealthCheckConfiguration (Maybe Natural)

The number of consecutive checks that must succeed before App Runner decides that the service is healthy.

Default: 1

healthCheckConfiguration_path :: Lens' HealthCheckConfiguration (Maybe Text)

The URL that health check requests are sent to.

Path is only applicable when you set Protocol to HTTP.

Default: "/"

healthCheckConfiguration_protocol :: Lens' HealthCheckConfiguration (Maybe HealthCheckProtocol)

The IP protocol that App Runner uses to perform health checks for your service.

If you set Protocol to HTTP, App Runner sends health check requests to the HTTP path specified by Path.

Default: TCP

healthCheckConfiguration_interval :: Lens' HealthCheckConfiguration (Maybe Natural)

The time interval, in seconds, between health checks.

Default: 5

healthCheckConfiguration_timeout :: Lens' HealthCheckConfiguration (Maybe Natural)

The time, in seconds, to wait for a health check response before deciding it failed.

Default: 2

healthCheckConfiguration_unhealthyThreshold :: Lens' HealthCheckConfiguration (Maybe Natural)

The number of consecutive checks that must fail before App Runner decides that the service is unhealthy.

Default: 5