Amazonka.AppRunner.Types.ImageRepository

Description

Documentation

data ImageRepository

Describes a source image repository.

See: newImageRepository smart constructor.

Constructors

ImageRepository'

Fields

imageConfiguration :: Maybe ImageConfiguration

Configuration for running the identified image.

imageIdentifier :: Text

The identifier of an image.

For an image in Amazon Elastic Container Registry (Amazon ECR), this is an image name. For the image name format, see Pulling an image in the Amazon ECR User Guide.

imageRepositoryType :: ImageRepositoryType

The type of the image repository. This reflects the repository provider and whether the repository is private or public.

newImageRepository

Arguments

:: Text
$sel:imageIdentifier:ImageRepository'

-> ImageRepositoryType
$sel:imageRepositoryType:ImageRepository'

-> ImageRepository

Create a value of ImageRepository with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:imageConfiguration:ImageRepository', imageRepository_imageConfiguration - Configuration for running the identified image.

$sel:imageIdentifier:ImageRepository', imageRepository_imageIdentifier - The identifier of an image.

For an image in Amazon Elastic Container Registry (Amazon ECR), this is an image name. For the image name format, see Pulling an image in the Amazon ECR User Guide.

$sel:imageRepositoryType:ImageRepository', imageRepository_imageRepositoryType - The type of the image repository. This reflects the repository provider and whether the repository is private or public.

imageRepository_imageConfiguration :: Lens' ImageRepository (Maybe ImageConfiguration)

Configuration for running the identified image.

imageRepository_imageIdentifier :: Lens' ImageRepository Text

The identifier of an image.

For an image in Amazon Elastic Container Registry (Amazon ECR), this is an image name. For the image name format, see Pulling an image in the Amazon ECR User Guide.

imageRepository_imageRepositoryType :: Lens' ImageRepository ImageRepositoryType

The type of the image repository. This reflects the repository provider and whether the repository is private or public.