Amazonka.AppRunner.Types.InstanceConfiguration

Description

Documentation

data InstanceConfiguration

Describes the runtime configuration of an App Runner service instance (scaling unit).

See: newInstanceConfiguration smart constructor.

Constructors

InstanceConfiguration'

Fields

memory :: Maybe Text

The amount of memory, in MB or GB, reserved for each instance of your App Runner service.

Default: 2 GB

instanceRoleArn :: Maybe Text

The Amazon Resource Name (ARN) of an IAM role that provides permissions to your App Runner service. These are permissions that your code needs when it calls any Amazon Web Services APIs.

cpu :: Maybe Text

The number of CPU units reserved for each instance of your App Runner service.

Default: 1 vCPU

newInstanceConfiguration :: InstanceConfiguration

Create a value of InstanceConfiguration with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:memory:InstanceConfiguration', instanceConfiguration_memory - The amount of memory, in MB or GB, reserved for each instance of your App Runner service.

Default: 2 GB

$sel:instanceRoleArn:InstanceConfiguration', instanceConfiguration_instanceRoleArn - The Amazon Resource Name (ARN) of an IAM role that provides permissions to your App Runner service. These are permissions that your code needs when it calls any Amazon Web Services APIs.

$sel:cpu:InstanceConfiguration', instanceConfiguration_cpu - The number of CPU units reserved for each instance of your App Runner service.

Default: 1 vCPU

instanceConfiguration_memory :: Lens' InstanceConfiguration (Maybe Text)

The amount of memory, in MB or GB, reserved for each instance of your App Runner service.

Default: 2 GB

instanceConfiguration_instanceRoleArn :: Lens' InstanceConfiguration (Maybe Text)

The Amazon Resource Name (ARN) of an IAM role that provides permissions to your App Runner service. These are permissions that your code needs when it calls any Amazon Web Services APIs.

instanceConfiguration_cpu :: Lens' InstanceConfiguration (Maybe Text)

The number of CPU units reserved for each instance of your App Runner service.

Default: 1 vCPU