Amazonka.AppRunner.Types.Service

Description

Documentation

data Service

Describes an App Runner service. It can describe a service in any state, including deleted services.

This type contains the full information about a service, including configuration details. It's returned by the CreateService, DescribeService, and DeleteService actions. A subset of this information is returned by the ListServices action using the ServiceSummary type.

See: newService smart constructor.

Constructors

Service'

Fields

encryptionConfiguration :: Maybe EncryptionConfiguration

The encryption key that App Runner uses to encrypt the service logs and the copy of the source repository that App Runner maintains for the service. It can be either a customer-provided encryption key or an Amazon Web Services managed CMK.

healthCheckConfiguration :: Maybe HealthCheckConfiguration

The settings for the health check that App Runner performs to monitor the health of this service.

deletedAt :: Maybe POSIX

The time when the App Runner service was deleted. It's in the Unix time stamp format.

serviceName :: Text

The customer-provided service name.

serviceId :: Text

An ID that App Runner generated for this service. It's unique within the Amazon Web Services Region.

serviceArn :: Text

The Amazon Resource Name (ARN) of this service.

serviceUrl :: Text

A subdomain URL that App Runner generated for this service. You can use this URL to access your service web application.

createdAt :: POSIX

The time when the App Runner service was created. It's in the Unix time stamp format.

updatedAt :: POSIX

The time when the App Runner service was last updated at. It's in the Unix time stamp format.

status :: ServiceStatus

The current state of the App Runner service. These particular values mean the following.

CREATE_FAILED – The service failed to create. To troubleshoot this failure, read the failure events and logs, change any parameters that need to be fixed, and retry the call to create the service.

The failed service isn't usable, and still counts towards your service quota. When you're done analyzing the failure, delete the service.

DELETE_FAILED – The service failed to delete and can't be successfully recovered. Retry the service deletion call to ensure that all related resources are removed.

sourceConfiguration :: SourceConfiguration

The source deployed to the App Runner service. It can be a code or an image repository.

instanceConfiguration :: InstanceConfiguration

The runtime configuration of instances (scaling units) of this service.

autoScalingConfigurationSummary :: AutoScalingConfigurationSummary

Summary information for the App Runner automatic scaling configuration resource that's associated with this service.

newService

Arguments

:: Text
$sel:serviceName:Service'

-> Text
$sel:serviceId:Service'

-> Text
$sel:serviceArn:Service'

-> Text
$sel:serviceUrl:Service'

-> UTCTime
$sel:createdAt:Service'

-> UTCTime
$sel:updatedAt:Service'

-> ServiceStatus
$sel:status:Service'

-> SourceConfiguration
$sel:sourceConfiguration:Service'

-> InstanceConfiguration
$sel:instanceConfiguration:Service'

-> AutoScalingConfigurationSummary
$sel:autoScalingConfigurationSummary:Service'

-> Service

Create a value of Service with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:encryptionConfiguration:Service', service_encryptionConfiguration - The encryption key that App Runner uses to encrypt the service logs and the copy of the source repository that App Runner maintains for the service. It can be either a customer-provided encryption key or an Amazon Web Services managed CMK.

$sel:healthCheckConfiguration:Service', service_healthCheckConfiguration - The settings for the health check that App Runner performs to monitor the health of this service.

$sel:deletedAt:Service', service_deletedAt - The time when the App Runner service was deleted. It's in the Unix time stamp format.

$sel:serviceName:Service', service_serviceName - The customer-provided service name.

$sel:serviceId:Service', service_serviceId - An ID that App Runner generated for this service. It's unique within the Amazon Web Services Region.

$sel:serviceArn:Service', service_serviceArn - The Amazon Resource Name (ARN) of this service.

$sel:serviceUrl:Service', service_serviceUrl - A subdomain URL that App Runner generated for this service. You can use this URL to access your service web application.

$sel:createdAt:Service', service_createdAt - The time when the App Runner service was created. It's in the Unix time stamp format.

$sel:updatedAt:Service', service_updatedAt - The time when the App Runner service was last updated at. It's in the Unix time stamp format.

$sel:status:Service', service_status - The current state of the App Runner service. These particular values mean the following.

CREATE_FAILED – The service failed to create. To troubleshoot this failure, read the failure events and logs, change any parameters that need to be fixed, and retry the call to create the service.

The failed service isn't usable, and still counts towards your service quota. When you're done analyzing the failure, delete the service.

DELETE_FAILED – The service failed to delete and can't be successfully recovered. Retry the service deletion call to ensure that all related resources are removed.

$sel:sourceConfiguration:Service', service_sourceConfiguration - The source deployed to the App Runner service. It can be a code or an image repository.

$sel:instanceConfiguration:Service', service_instanceConfiguration - The runtime configuration of instances (scaling units) of this service.

$sel:autoScalingConfigurationSummary:Service', service_autoScalingConfigurationSummary - Summary information for the App Runner automatic scaling configuration resource that's associated with this service.

service_encryptionConfiguration :: Lens' Service (Maybe EncryptionConfiguration)

The encryption key that App Runner uses to encrypt the service logs and the copy of the source repository that App Runner maintains for the service. It can be either a customer-provided encryption key or an Amazon Web Services managed CMK.

service_healthCheckConfiguration :: Lens' Service (Maybe HealthCheckConfiguration)

The settings for the health check that App Runner performs to monitor the health of this service.

service_deletedAt :: Lens' Service (Maybe UTCTime)

The time when the App Runner service was deleted. It's in the Unix time stamp format.

service_serviceName :: Lens' Service Text

The customer-provided service name.

service_serviceId :: Lens' Service Text

An ID that App Runner generated for this service. It's unique within the Amazon Web Services Region.

service_serviceArn :: Lens' Service Text

The Amazon Resource Name (ARN) of this service.

service_serviceUrl :: Lens' Service Text

A subdomain URL that App Runner generated for this service. You can use this URL to access your service web application.

service_createdAt :: Lens' Service UTCTime

The time when the App Runner service was created. It's in the Unix time stamp format.

service_updatedAt :: Lens' Service UTCTime

The time when the App Runner service was last updated at. It's in the Unix time stamp format.

service_status :: Lens' Service ServiceStatus

The current state of the App Runner service. These particular values mean the following.

CREATE_FAILED – The service failed to create. To troubleshoot this failure, read the failure events and logs, change any parameters that need to be fixed, and retry the call to create the service.

The failed service isn't usable, and still counts towards your service quota. When you're done analyzing the failure, delete the service.

DELETE_FAILED – The service failed to delete and can't be successfully recovered. Retry the service deletion call to ensure that all related resources are removed.

service_sourceConfiguration :: Lens' Service SourceConfiguration

The source deployed to the App Runner service. It can be a code or an image repository.

service_instanceConfiguration :: Lens' Service InstanceConfiguration

The runtime configuration of instances (scaling units) of this service.

service_autoScalingConfigurationSummary :: Lens' Service AutoScalingConfigurationSummary

Summary information for the App Runner automatic scaling configuration resource that's associated with this service.