Amazonka.AppStream.Types.AccessEndpoint

Description

Documentation

data AccessEndpoint

Describes an interface VPC endpoint (interface endpoint) that lets you create a private connection between the virtual private cloud (VPC) that you specify and AppStream 2.0. When you specify an interface endpoint for a stack, users of the stack can connect to AppStream 2.0 only through that endpoint. When you specify an interface endpoint for an image builder, administrators can connect to the image builder only through that endpoint.

See: newAccessEndpoint smart constructor.

Constructors

AccessEndpoint'

Fields

vpceId :: Maybe Text

The identifier (ID) of the VPC in which the interface endpoint is used.

endpointType :: AccessEndpointType

The type of interface endpoint.

newAccessEndpoint

Arguments

:: AccessEndpointType
$sel:endpointType:AccessEndpoint'

-> AccessEndpoint

Create a value of AccessEndpoint with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:vpceId:AccessEndpoint', accessEndpoint_vpceId - The identifier (ID) of the VPC in which the interface endpoint is used.

$sel:endpointType:AccessEndpoint', accessEndpoint_endpointType - The type of interface endpoint.

accessEndpoint_vpceId :: Lens' AccessEndpoint (Maybe Text)

The identifier (ID) of the VPC in which the interface endpoint is used.

accessEndpoint_endpointType :: Lens' AccessEndpoint AccessEndpointType

The type of interface endpoint.