Amazonka.AppStream.Types.Application

Description

Documentation

data Application

Describes an application in the application catalog.

See: newApplication smart constructor.

Constructors

Application'

Fields

enabled :: Maybe Bool

If there is a problem, the application can be disabled after image creation.

launchPath :: Maybe Text

The path to the application executable in the instance.

launchParameters :: Maybe Text

The arguments that are passed to the application at launch.

name :: Maybe Text

The name of the application.

displayName :: Maybe Text

The application name to display.

metadata :: Maybe (HashMap Text Text)

Additional attributes that describe the application.

iconURL :: Maybe Text

The URL for the application icon. This URL might be time-limited.

newApplication :: Application

Create a value of Application with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:enabled:Application', application_enabled - If there is a problem, the application can be disabled after image creation.

$sel:launchPath:Application', application_launchPath - The path to the application executable in the instance.

$sel:launchParameters:Application', application_launchParameters - The arguments that are passed to the application at launch.

$sel:name:Application', application_name - The name of the application.

$sel:displayName:Application', application_displayName - The application name to display.

$sel:metadata:Application', application_metadata - Additional attributes that describe the application.

$sel:iconURL:Application', application_iconURL - The URL for the application icon. This URL might be time-limited.

application_enabled :: Lens' Application (Maybe Bool)

If there is a problem, the application can be disabled after image creation.

application_launchPath :: Lens' Application (Maybe Text)

The path to the application executable in the instance.

application_launchParameters :: Lens' Application (Maybe Text)

The arguments that are passed to the application at launch.

application_name :: Lens' Application (Maybe Text)

The name of the application.

application_displayName :: Lens' Application (Maybe Text)

The application name to display.

application_metadata :: Lens' Application (Maybe (HashMap Text Text))

Additional attributes that describe the application.

application_iconURL :: Lens' Application (Maybe Text)

The URL for the application icon. This URL might be time-limited.