Amazonka.AppStream.Types.ApplicationSettings

Description

Documentation

data ApplicationSettings

The persistent application settings for users of a stack.

See: newApplicationSettings smart constructor.

Constructors

ApplicationSettings'

Fields

settingsGroup :: Maybe Text

The path prefix for the S3 bucket where users’ persistent application settings are stored. You can allow the same persistent application settings to be used across multiple stacks by specifying the same settings group for each stack.

enabled :: Bool

Enables or disables persistent application settings for users during their streaming sessions.

newApplicationSettings

Arguments

:: Bool
$sel:enabled:ApplicationSettings'

-> ApplicationSettings

Create a value of ApplicationSettings with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:settingsGroup:ApplicationSettings', applicationSettings_settingsGroup - The path prefix for the S3 bucket where users’ persistent application settings are stored. You can allow the same persistent application settings to be used across multiple stacks by specifying the same settings group for each stack.

$sel:enabled:ApplicationSettings', applicationSettings_enabled - Enables or disables persistent application settings for users during their streaming sessions.

applicationSettings_settingsGroup :: Lens' ApplicationSettings (Maybe Text)

The path prefix for the S3 bucket where users’ persistent application settings are stored. You can allow the same persistent application settings to be used across multiple stacks by specifying the same settings group for each stack.

applicationSettings_enabled :: Lens' ApplicationSettings Bool

Enables or disables persistent application settings for users during their streaming sessions.