Amazonka.AppStream.Types.DirectoryConfig

Description

Documentation

data DirectoryConfig

Describes the configuration information required to join fleets and image builders to Microsoft Active Directory domains.

See: newDirectoryConfig smart constructor.

Constructors

DirectoryConfig'

Fields

createdTime :: Maybe POSIX

The time the directory configuration was created.

serviceAccountCredentials :: Maybe ServiceAccountCredentials

The credentials for the service account used by the fleet or image builder to connect to the directory.

organizationalUnitDistinguishedNames :: Maybe [Text]

The distinguished names of the organizational units for computer accounts.

directoryName :: Text

The fully qualified name of the directory (for example, corp.example.com).

newDirectoryConfig

Arguments

:: Text
$sel:directoryName:DirectoryConfig'

-> DirectoryConfig

Create a value of DirectoryConfig with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:createdTime:DirectoryConfig', directoryConfig_createdTime - The time the directory configuration was created.

$sel:serviceAccountCredentials:DirectoryConfig', directoryConfig_serviceAccountCredentials - The credentials for the service account used by the fleet or image builder to connect to the directory.

$sel:organizationalUnitDistinguishedNames:DirectoryConfig', directoryConfig_organizationalUnitDistinguishedNames - The distinguished names of the organizational units for computer accounts.

$sel:directoryName:DirectoryConfig', directoryConfig_directoryName - The fully qualified name of the directory (for example, corp.example.com).

directoryConfig_createdTime :: Lens' DirectoryConfig (Maybe UTCTime)

The time the directory configuration was created.

directoryConfig_serviceAccountCredentials :: Lens' DirectoryConfig (Maybe ServiceAccountCredentials)

The credentials for the service account used by the fleet or image builder to connect to the directory.

directoryConfig_organizationalUnitDistinguishedNames :: Lens' DirectoryConfig (Maybe [Text])

The distinguished names of the organizational units for computer accounts.

directoryConfig_directoryName :: Lens' DirectoryConfig Text

The fully qualified name of the directory (for example, corp.example.com).