Amazonka.AppStream.Types.Image

Description

Documentation

data Image

Describes an image.

See: newImage smart constructor.

Constructors

Image'

Fields

state :: Maybe ImageState

The image starts in the PENDING state. If image creation succeeds, the state is AVAILABLE. If image creation fails, the state is FAILED.

imagePermissions :: Maybe ImagePermissions

The permissions to provide to the destination AWS account for the specified image.

platform :: Maybe PlatformType

The operating system platform of the image.

publicBaseImageReleasedDate :: Maybe POSIX

The release date of the public base image. For private images, this date is the release date of the base image from which the image was created.

stateChangeReason :: Maybe ImageStateChangeReason

The reason why the last state change occurred.

arn :: Maybe Text

The ARN of the image.

createdTime :: Maybe POSIX

The time the image was created.

imageBuilderSupported :: Maybe Bool

Indicates whether an image builder can be launched from this image.

visibility :: Maybe VisibilityType

Indicates whether the image is public or private.

imageBuilderName :: Maybe Text

The name of the image builder that was used to create the private image. If the image is shared, this value is null.

imageErrors :: Maybe [ResourceError]

Describes the errors that are returned when a new image can't be created.

baseImageArn :: Maybe Text

The ARN of the image from which this image was created.

displayName :: Maybe Text

The image name to display.

description :: Maybe Text

The description to display.

appstreamAgentVersion :: Maybe Text

The version of the AppStream 2.0 agent to use for instances that are launched from this image.

applications :: Maybe [Application]

The applications associated with the image.

name :: Text

The name of the image.

newImage

Arguments

:: Text
$sel:name:Image'

-> Image

Create a value of Image with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:state:Image', image_state - The image starts in the PENDING state. If image creation succeeds, the state is AVAILABLE. If image creation fails, the state is FAILED.

$sel:imagePermissions:Image', image_imagePermissions - The permissions to provide to the destination AWS account for the specified image.

$sel:platform:Image', image_platform - The operating system platform of the image.

$sel:publicBaseImageReleasedDate:Image', image_publicBaseImageReleasedDate - The release date of the public base image. For private images, this date is the release date of the base image from which the image was created.

$sel:stateChangeReason:Image', image_stateChangeReason - The reason why the last state change occurred.

$sel:arn:Image', image_arn - The ARN of the image.

$sel:createdTime:Image', image_createdTime - The time the image was created.

$sel:imageBuilderSupported:Image', image_imageBuilderSupported - Indicates whether an image builder can be launched from this image.

$sel:visibility:Image', image_visibility - Indicates whether the image is public or private.

$sel:imageBuilderName:Image', image_imageBuilderName - The name of the image builder that was used to create the private image. If the image is shared, this value is null.

$sel:imageErrors:Image', image_imageErrors - Describes the errors that are returned when a new image can't be created.

$sel:baseImageArn:Image', image_baseImageArn - The ARN of the image from which this image was created.

$sel:displayName:Image', image_displayName - The image name to display.

$sel:description:Image', image_description - The description to display.

$sel:appstreamAgentVersion:Image', image_appstreamAgentVersion - The version of the AppStream 2.0 agent to use for instances that are launched from this image.

$sel:applications:Image', image_applications - The applications associated with the image.

$sel:name:Image', image_name - The name of the image.

image_state :: Lens' Image (Maybe ImageState)

The image starts in the PENDING state. If image creation succeeds, the state is AVAILABLE. If image creation fails, the state is FAILED.

image_imagePermissions :: Lens' Image (Maybe ImagePermissions)

The permissions to provide to the destination AWS account for the specified image.

image_platform :: Lens' Image (Maybe PlatformType)

The operating system platform of the image.

image_publicBaseImageReleasedDate :: Lens' Image (Maybe UTCTime)

The release date of the public base image. For private images, this date is the release date of the base image from which the image was created.

image_stateChangeReason :: Lens' Image (Maybe ImageStateChangeReason)

The reason why the last state change occurred.

image_arn :: Lens' Image (Maybe Text)

The ARN of the image.

image_createdTime :: Lens' Image (Maybe UTCTime)

The time the image was created.

image_imageBuilderSupported :: Lens' Image (Maybe Bool)

Indicates whether an image builder can be launched from this image.

image_visibility :: Lens' Image (Maybe VisibilityType)

Indicates whether the image is public or private.

image_imageBuilderName :: Lens' Image (Maybe Text)

The name of the image builder that was used to create the private image. If the image is shared, this value is null.

image_imageErrors :: Lens' Image (Maybe [ResourceError])

Describes the errors that are returned when a new image can't be created.

image_baseImageArn :: Lens' Image (Maybe Text)

The ARN of the image from which this image was created.

image_displayName :: Lens' Image (Maybe Text)

The image name to display.

image_description :: Lens' Image (Maybe Text)

The description to display.

image_appstreamAgentVersion :: Lens' Image (Maybe Text)

The version of the AppStream 2.0 agent to use for instances that are launched from this image.

image_applications :: Lens' Image (Maybe [Application])

The applications associated with the image.

image_name :: Lens' Image Text

The name of the image.