Amazonka.AppStream.Types.Stack

Description

Documentation

data Stack

Describes a stack.

See: newStack smart constructor.

Constructors

Stack'

Fields

userSettings :: Maybe (NonEmpty UserSetting)

The actions that are enabled or disabled for users during their streaming sessions. By default these actions are enabled.

applicationSettings :: Maybe ApplicationSettingsResponse

The persistent application settings for users of the stack.

feedbackURL :: Maybe Text

The URL that users are redirected to after they click the Send Feedback link. If no URL is specified, no Send Feedback link is displayed.

arn :: Maybe Text

The ARN of the stack.

createdTime :: Maybe POSIX

The time the stack was created.

storageConnectors :: Maybe [StorageConnector]

The storage connectors to enable.

accessEndpoints :: Maybe (NonEmpty AccessEndpoint)

The list of virtual private cloud (VPC) interface endpoint objects. Users of the stack can connect to AppStream 2.0 only through the specified endpoints.

displayName :: Maybe Text

The stack name to display.

stackErrors :: Maybe [StackError]

The errors for the stack.

embedHostDomains :: Maybe (NonEmpty Text)

The domains where AppStream 2.0 streaming sessions can be embedded in an iframe. You must approve the domains that you want to host embedded AppStream 2.0 streaming sessions.

description :: Maybe Text

The description to display.

redirectURL :: Maybe Text

The URL that users are redirected to after their streaming session ends.

name :: Text

The name of the stack.

newStack

Arguments

:: Text
$sel:name:Stack'

-> Stack

Create a value of Stack with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:userSettings:Stack', stack_userSettings - The actions that are enabled or disabled for users during their streaming sessions. By default these actions are enabled.

$sel:applicationSettings:Stack', stack_applicationSettings - The persistent application settings for users of the stack.

$sel:feedbackURL:Stack', stack_feedbackURL - The URL that users are redirected to after they click the Send Feedback link. If no URL is specified, no Send Feedback link is displayed.

$sel:arn:Stack', stack_arn - The ARN of the stack.

$sel:createdTime:Stack', stack_createdTime - The time the stack was created.

$sel:storageConnectors:Stack', stack_storageConnectors - The storage connectors to enable.

$sel:accessEndpoints:Stack', stack_accessEndpoints - The list of virtual private cloud (VPC) interface endpoint objects. Users of the stack can connect to AppStream 2.0 only through the specified endpoints.

$sel:displayName:Stack', stack_displayName - The stack name to display.

$sel:stackErrors:Stack', stack_stackErrors - The errors for the stack.

$sel:embedHostDomains:Stack', stack_embedHostDomains - The domains where AppStream 2.0 streaming sessions can be embedded in an iframe. You must approve the domains that you want to host embedded AppStream 2.0 streaming sessions.

$sel:description:Stack', stack_description - The description to display.

$sel:redirectURL:Stack', stack_redirectURL - The URL that users are redirected to after their streaming session ends.

$sel:name:Stack', stack_name - The name of the stack.

stack_userSettings :: Lens' Stack (Maybe (NonEmpty UserSetting))

The actions that are enabled or disabled for users during their streaming sessions. By default these actions are enabled.

stack_applicationSettings :: Lens' Stack (Maybe ApplicationSettingsResponse)

The persistent application settings for users of the stack.

stack_feedbackURL :: Lens' Stack (Maybe Text)

The URL that users are redirected to after they click the Send Feedback link. If no URL is specified, no Send Feedback link is displayed.

stack_arn :: Lens' Stack (Maybe Text)

The ARN of the stack.

stack_createdTime :: Lens' Stack (Maybe UTCTime)

The time the stack was created.

stack_storageConnectors :: Lens' Stack (Maybe [StorageConnector])

The storage connectors to enable.

stack_accessEndpoints :: Lens' Stack (Maybe (NonEmpty AccessEndpoint))

The list of virtual private cloud (VPC) interface endpoint objects. Users of the stack can connect to AppStream 2.0 only through the specified endpoints.

stack_displayName :: Lens' Stack (Maybe Text)

The stack name to display.

stack_stackErrors :: Lens' Stack (Maybe [StackError])

The errors for the stack.

stack_embedHostDomains :: Lens' Stack (Maybe (NonEmpty Text))

The domains where AppStream 2.0 streaming sessions can be embedded in an iframe. You must approve the domains that you want to host embedded AppStream 2.0 streaming sessions.

stack_description :: Lens' Stack (Maybe Text)

The description to display.

stack_redirectURL :: Lens' Stack (Maybe Text)

The URL that users are redirected to after their streaming session ends.

stack_name :: Lens' Stack Text

The name of the stack.