Amazonka.AppStream.UpdateFleet

Description

Updates the specified fleet.

If the fleet is in the STOPPED state, you can update any attribute except the fleet name. If the fleet is in the RUNNING state, you can update the DisplayName, ComputeCapacity, ImageARN, ImageName, IdleDisconnectTimeoutInSeconds, and DisconnectTimeoutInSeconds attributes. If the fleet is in the STARTING or STOPPING state, you can't update it.

Creating a Request

data UpdateFleet

See: newUpdateFleet smart constructor.

Constructors

UpdateFleet'

Fields

domainJoinInfo :: Maybe DomainJoinInfo

The name of the directory and organizational unit (OU) to use to join the fleet to a Microsoft Active Directory domain.

iamRoleArn :: Maybe Text

The Amazon Resource Name (ARN) of the IAM role to apply to the fleet. To assume a role, a fleet instance calls the AWS Security Token Service (STS) AssumeRole API operation and passes the ARN of the role to use. The operation creates a new session with temporary credentials. AppStream 2.0 retrieves the temporary credentials and creates the appstream_machine_role credential profile on the instance.

For more information, see Using an IAM Role to Grant Permissions to Applications and Scripts Running on AppStream 2.0 Streaming Instances in the Amazon AppStream 2.0 Administration Guide.

disconnectTimeoutInSeconds :: Maybe Int

The amount of time that a streaming session remains active after users disconnect. If users try to reconnect to the streaming session after a disconnection or network interruption within this time interval, they are connected to their previous session. Otherwise, they are connected to a new session with a new streaming instance.

Specify a value between 60 and 360000.

maxUserDurationInSeconds :: Maybe Int

The maximum amount of time that a streaming session can remain active, in seconds. If users are still connected to a streaming instance five minutes before this limit is reached, they are prompted to save any open documents before being disconnected. After this time elapses, the instance is terminated and replaced by a new instance.

Specify a value between 600 and 360000.

attributesToDelete :: Maybe [FleetAttribute]

The fleet attributes to delete.

idleDisconnectTimeoutInSeconds :: Maybe Int

The amount of time that users can be idle (inactive) before they are disconnected from their streaming session and the DisconnectTimeoutInSeconds time interval begins. Users are notified before they are disconnected due to inactivity. If users try to reconnect to the streaming session before the time interval specified in DisconnectTimeoutInSeconds elapses, they are connected to their previous session. Users are considered idle when they stop providing keyboard or mouse input during their streaming session. File uploads and downloads, audio in, audio out, and pixels changing do not qualify as user activity. If users continue to be idle after the time interval in IdleDisconnectTimeoutInSeconds elapses, they are disconnected.

To prevent users from being disconnected due to inactivity, specify a value of 0. Otherwise, specify a value between 60 and 3600. The default value is 0.

If you enable this feature, we recommend that you specify a value that corresponds exactly to a whole number of minutes (for example, 60, 120, and 180). If you don't do this, the value is rounded to the nearest minute. For example, if you specify a value of 70, users are disconnected after 1 minute of inactivity. If you specify a value that is at the midpoint between two different minutes, the value is rounded up. For example, if you specify a value of 90, users are disconnected after 2 minutes of inactivity.

deleteVpcConfig :: Maybe Bool

Deletes the VPC association for the specified fleet.

instanceType :: Maybe Text

The instance type to use when launching fleet instances. The following instance types are available:

stream.standard.small

stream.standard.medium

stream.standard.large

stream.compute.large

stream.compute.xlarge

stream.compute.2xlarge

stream.compute.4xlarge

stream.compute.8xlarge

stream.memory.large

stream.memory.xlarge

stream.memory.2xlarge

stream.memory.4xlarge

stream.memory.8xlarge

stream.memory.z1d.large

stream.memory.z1d.xlarge

stream.memory.z1d.2xlarge

stream.memory.z1d.3xlarge

stream.memory.z1d.6xlarge

stream.memory.z1d.12xlarge

stream.graphics-design.large

stream.graphics-design.xlarge

stream.graphics-design.2xlarge

stream.graphics-design.4xlarge

stream.graphics-desktop.2xlarge

stream.graphics.g4dn.xlarge

stream.graphics.g4dn.2xlarge

stream.graphics.g4dn.4xlarge

stream.graphics.g4dn.8xlarge

stream.graphics.g4dn.12xlarge

stream.graphics.g4dn.16xlarge

stream.graphics-pro.4xlarge

stream.graphics-pro.8xlarge

stream.graphics-pro.16xlarge

vpcConfig :: Maybe VpcConfig

The VPC configuration for the fleet.

name :: Maybe Text

A unique name for the fleet.

imageArn :: Maybe Text

The ARN of the public, private, or shared image to use.

displayName :: Maybe Text

The fleet name to display.

enableDefaultInternetAccess :: Maybe Bool

Enables or disables default internet access for the fleet.

imageName :: Maybe Text

The name of the image used to create the fleet.

description :: Maybe Text

The description to display.

streamView :: Maybe StreamView

The AppStream 2.0 view that is displayed to your users when they stream from the fleet. When APP is specified, only the windows of applications opened by users display. When DESKTOP is specified, the standard desktop that is provided by the operating system displays.

The default value is APP.

computeCapacity :: Maybe ComputeCapacity

The desired capacity for the fleet.

newUpdateFleet :: UpdateFleet

Create a value of UpdateFleet with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:domainJoinInfo:UpdateFleet', updateFleet_domainJoinInfo - The name of the directory and organizational unit (OU) to use to join the fleet to a Microsoft Active Directory domain.

$sel:iamRoleArn:UpdateFleet', updateFleet_iamRoleArn - The Amazon Resource Name (ARN) of the IAM role to apply to the fleet. To assume a role, a fleet instance calls the AWS Security Token Service (STS) AssumeRole API operation and passes the ARN of the role to use. The operation creates a new session with temporary credentials. AppStream 2.0 retrieves the temporary credentials and creates the appstream_machine_role credential profile on the instance.

For more information, see Using an IAM Role to Grant Permissions to Applications and Scripts Running on AppStream 2.0 Streaming Instances in the Amazon AppStream 2.0 Administration Guide.

$sel:disconnectTimeoutInSeconds:UpdateFleet', updateFleet_disconnectTimeoutInSeconds - The amount of time that a streaming session remains active after users disconnect. If users try to reconnect to the streaming session after a disconnection or network interruption within this time interval, they are connected to their previous session. Otherwise, they are connected to a new session with a new streaming instance.

Specify a value between 60 and 360000.

$sel:maxUserDurationInSeconds:UpdateFleet', updateFleet_maxUserDurationInSeconds - The maximum amount of time that a streaming session can remain active, in seconds. If users are still connected to a streaming instance five minutes before this limit is reached, they are prompted to save any open documents before being disconnected. After this time elapses, the instance is terminated and replaced by a new instance.

Specify a value between 600 and 360000.

$sel:attributesToDelete:UpdateFleet', updateFleet_attributesToDelete - The fleet attributes to delete.

$sel:idleDisconnectTimeoutInSeconds:UpdateFleet', updateFleet_idleDisconnectTimeoutInSeconds - The amount of time that users can be idle (inactive) before they are disconnected from their streaming session and the DisconnectTimeoutInSeconds time interval begins. Users are notified before they are disconnected due to inactivity. If users try to reconnect to the streaming session before the time interval specified in DisconnectTimeoutInSeconds elapses, they are connected to their previous session. Users are considered idle when they stop providing keyboard or mouse input during their streaming session. File uploads and downloads, audio in, audio out, and pixels changing do not qualify as user activity. If users continue to be idle after the time interval in IdleDisconnectTimeoutInSeconds elapses, they are disconnected.

To prevent users from being disconnected due to inactivity, specify a value of 0. Otherwise, specify a value between 60 and 3600. The default value is 0.

If you enable this feature, we recommend that you specify a value that corresponds exactly to a whole number of minutes (for example, 60, 120, and 180). If you don't do this, the value is rounded to the nearest minute. For example, if you specify a value of 70, users are disconnected after 1 minute of inactivity. If you specify a value that is at the midpoint between two different minutes, the value is rounded up. For example, if you specify a value of 90, users are disconnected after 2 minutes of inactivity.

$sel:deleteVpcConfig:UpdateFleet', updateFleet_deleteVpcConfig - Deletes the VPC association for the specified fleet.

$sel:instanceType:UpdateFleet', updateFleet_instanceType - The instance type to use when launching fleet instances. The following instance types are available:

stream.standard.small

stream.standard.medium

stream.standard.large

stream.compute.large

stream.compute.xlarge

stream.compute.2xlarge

stream.compute.4xlarge

stream.compute.8xlarge

stream.memory.large

stream.memory.xlarge

stream.memory.2xlarge

stream.memory.4xlarge

stream.memory.8xlarge

stream.memory.z1d.large

stream.memory.z1d.xlarge

stream.memory.z1d.2xlarge

stream.memory.z1d.3xlarge

stream.memory.z1d.6xlarge

stream.memory.z1d.12xlarge

stream.graphics-design.large

stream.graphics-design.xlarge

stream.graphics-design.2xlarge

stream.graphics-design.4xlarge

stream.graphics-desktop.2xlarge

stream.graphics.g4dn.xlarge

stream.graphics.g4dn.2xlarge

stream.graphics.g4dn.4xlarge

stream.graphics.g4dn.8xlarge

stream.graphics.g4dn.12xlarge

stream.graphics.g4dn.16xlarge

stream.graphics-pro.4xlarge

stream.graphics-pro.8xlarge

stream.graphics-pro.16xlarge

$sel:vpcConfig:UpdateFleet', updateFleet_vpcConfig - The VPC configuration for the fleet.

$sel:name:UpdateFleet', updateFleet_name - A unique name for the fleet.

$sel:imageArn:UpdateFleet', updateFleet_imageArn - The ARN of the public, private, or shared image to use.

$sel:displayName:UpdateFleet', updateFleet_displayName - The fleet name to display.

$sel:enableDefaultInternetAccess:UpdateFleet', updateFleet_enableDefaultInternetAccess - Enables or disables default internet access for the fleet.

$sel:imageName:UpdateFleet', updateFleet_imageName - The name of the image used to create the fleet.

$sel:description:UpdateFleet', updateFleet_description - The description to display.

$sel:streamView:UpdateFleet', updateFleet_streamView - The AppStream 2.0 view that is displayed to your users when they stream from the fleet. When APP is specified, only the windows of applications opened by users display. When DESKTOP is specified, the standard desktop that is provided by the operating system displays.

The default value is APP.

$sel:computeCapacity:UpdateFleet', updateFleet_computeCapacity - The desired capacity for the fleet.

Request Lenses

updateFleet_domainJoinInfo :: Lens' UpdateFleet (Maybe DomainJoinInfo)

The name of the directory and organizational unit (OU) to use to join the fleet to a Microsoft Active Directory domain.

updateFleet_iamRoleArn :: Lens' UpdateFleet (Maybe Text)

The Amazon Resource Name (ARN) of the IAM role to apply to the fleet. To assume a role, a fleet instance calls the AWS Security Token Service (STS) AssumeRole API operation and passes the ARN of the role to use. The operation creates a new session with temporary credentials. AppStream 2.0 retrieves the temporary credentials and creates the appstream_machine_role credential profile on the instance.

For more information, see Using an IAM Role to Grant Permissions to Applications and Scripts Running on AppStream 2.0 Streaming Instances in the Amazon AppStream 2.0 Administration Guide.

updateFleet_disconnectTimeoutInSeconds :: Lens' UpdateFleet (Maybe Int)

The amount of time that a streaming session remains active after users disconnect. If users try to reconnect to the streaming session after a disconnection or network interruption within this time interval, they are connected to their previous session. Otherwise, they are connected to a new session with a new streaming instance.

Specify a value between 60 and 360000.

updateFleet_maxUserDurationInSeconds :: Lens' UpdateFleet (Maybe Int)

The maximum amount of time that a streaming session can remain active, in seconds. If users are still connected to a streaming instance five minutes before this limit is reached, they are prompted to save any open documents before being disconnected. After this time elapses, the instance is terminated and replaced by a new instance.

Specify a value between 600 and 360000.

updateFleet_attributesToDelete :: Lens' UpdateFleet (Maybe [FleetAttribute])

The fleet attributes to delete.

updateFleet_idleDisconnectTimeoutInSeconds :: Lens' UpdateFleet (Maybe Int)

The amount of time that users can be idle (inactive) before they are disconnected from their streaming session and the DisconnectTimeoutInSeconds time interval begins. Users are notified before they are disconnected due to inactivity. If users try to reconnect to the streaming session before the time interval specified in DisconnectTimeoutInSeconds elapses, they are connected to their previous session. Users are considered idle when they stop providing keyboard or mouse input during their streaming session. File uploads and downloads, audio in, audio out, and pixels changing do not qualify as user activity. If users continue to be idle after the time interval in IdleDisconnectTimeoutInSeconds elapses, they are disconnected.

To prevent users from being disconnected due to inactivity, specify a value of 0. Otherwise, specify a value between 60 and 3600. The default value is 0.

If you enable this feature, we recommend that you specify a value that corresponds exactly to a whole number of minutes (for example, 60, 120, and 180). If you don't do this, the value is rounded to the nearest minute. For example, if you specify a value of 70, users are disconnected after 1 minute of inactivity. If you specify a value that is at the midpoint between two different minutes, the value is rounded up. For example, if you specify a value of 90, users are disconnected after 2 minutes of inactivity.

updateFleet_deleteVpcConfig :: Lens' UpdateFleet (Maybe Bool)

Deletes the VPC association for the specified fleet.

updateFleet_instanceType :: Lens' UpdateFleet (Maybe Text)

The instance type to use when launching fleet instances. The following instance types are available:

stream.standard.small

stream.standard.medium

stream.standard.large

stream.compute.large

stream.compute.xlarge

stream.compute.2xlarge

stream.compute.4xlarge

stream.compute.8xlarge

stream.memory.large

stream.memory.xlarge

stream.memory.2xlarge

stream.memory.4xlarge

stream.memory.8xlarge

stream.memory.z1d.large

stream.memory.z1d.xlarge

stream.memory.z1d.2xlarge

stream.memory.z1d.3xlarge

stream.memory.z1d.6xlarge

stream.memory.z1d.12xlarge

stream.graphics-design.large

stream.graphics-design.xlarge

stream.graphics-design.2xlarge

stream.graphics-design.4xlarge

stream.graphics-desktop.2xlarge

stream.graphics.g4dn.xlarge

stream.graphics.g4dn.2xlarge

stream.graphics.g4dn.4xlarge

stream.graphics.g4dn.8xlarge

stream.graphics.g4dn.12xlarge

stream.graphics.g4dn.16xlarge

stream.graphics-pro.4xlarge

stream.graphics-pro.8xlarge

stream.graphics-pro.16xlarge

updateFleet_vpcConfig :: Lens' UpdateFleet (Maybe VpcConfig)

The VPC configuration for the fleet.

updateFleet_name :: Lens' UpdateFleet (Maybe Text)

A unique name for the fleet.

updateFleet_imageArn :: Lens' UpdateFleet (Maybe Text)

The ARN of the public, private, or shared image to use.

updateFleet_displayName :: Lens' UpdateFleet (Maybe Text)

The fleet name to display.

updateFleet_enableDefaultInternetAccess :: Lens' UpdateFleet (Maybe Bool)

Enables or disables default internet access for the fleet.

updateFleet_imageName :: Lens' UpdateFleet (Maybe Text)

The name of the image used to create the fleet.

updateFleet_description :: Lens' UpdateFleet (Maybe Text)

The description to display.

updateFleet_streamView :: Lens' UpdateFleet (Maybe StreamView)

The AppStream 2.0 view that is displayed to your users when they stream from the fleet. When APP is specified, only the windows of applications opened by users display. When DESKTOP is specified, the standard desktop that is provided by the operating system displays.

The default value is APP.

updateFleet_computeCapacity :: Lens' UpdateFleet (Maybe ComputeCapacity)

The desired capacity for the fleet.

Destructuring the Response

data UpdateFleetResponse

See: newUpdateFleetResponse smart constructor.

Constructors

UpdateFleetResponse'

Fields

fleet :: Maybe Fleet

Information about the fleet.

httpStatus :: Int

The response's http status code.

newUpdateFleetResponse

Arguments

:: Int
$sel:httpStatus:UpdateFleetResponse'

-> UpdateFleetResponse

Create a value of UpdateFleetResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:fleet:UpdateFleetResponse', updateFleetResponse_fleet - Information about the fleet.

$sel:httpStatus:UpdateFleetResponse', updateFleetResponse_httpStatus - The response's http status code.

Response Lenses

updateFleetResponse_fleet :: Lens' UpdateFleetResponse (Maybe Fleet)

Information about the fleet.

updateFleetResponse_httpStatus :: Lens' UpdateFleetResponse Int

The response's http status code.