Amazonka.AppSync.CreateApiKey

Description

Creates a unique key that you can distribute to clients who are executing your API.

Creating a Request

data CreateApiKey

See: newCreateApiKey smart constructor.

Constructors

CreateApiKey'

Fields

expires :: Maybe Integer

The time from creation time after which the API key expires. The date is represented as seconds since the epoch, rounded down to the nearest hour. The default value for this parameter is 7 days from creation time. For more information, see .

description :: Maybe Text

A description of the purpose of the API key.

apiId :: Text

The ID for your GraphQL API.

newCreateApiKey

Arguments

:: Text
$sel:apiId:CreateApiKey'

-> CreateApiKey

Create a value of CreateApiKey with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:expires:CreateApiKey', createApiKey_expires - The time from creation time after which the API key expires. The date is represented as seconds since the epoch, rounded down to the nearest hour. The default value for this parameter is 7 days from creation time. For more information, see .

$sel:description:CreateApiKey', createApiKey_description - A description of the purpose of the API key.

$sel:apiId:CreateApiKey', createApiKey_apiId - The ID for your GraphQL API.

Request Lenses

createApiKey_expires :: Lens' CreateApiKey (Maybe Integer)

The time from creation time after which the API key expires. The date is represented as seconds since the epoch, rounded down to the nearest hour. The default value for this parameter is 7 days from creation time. For more information, see .

createApiKey_description :: Lens' CreateApiKey (Maybe Text)

A description of the purpose of the API key.

createApiKey_apiId :: Lens' CreateApiKey Text

The ID for your GraphQL API.

Destructuring the Response

data CreateApiKeyResponse

See: newCreateApiKeyResponse smart constructor.

Constructors

CreateApiKeyResponse'

Fields

apiKey :: Maybe ApiKey

The API key.

httpStatus :: Int

The response's http status code.

newCreateApiKeyResponse

Arguments

:: Int
$sel:httpStatus:CreateApiKeyResponse'

-> CreateApiKeyResponse

Create a value of CreateApiKeyResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:apiKey:CreateApiKeyResponse', createApiKeyResponse_apiKey - The API key.

$sel:httpStatus:CreateApiKeyResponse', createApiKeyResponse_httpStatus - The response's http status code.

Response Lenses

createApiKeyResponse_apiKey :: Lens' CreateApiKeyResponse (Maybe ApiKey)

The API key.

createApiKeyResponse_httpStatus :: Lens' CreateApiKeyResponse Int

The response's http status code.