Amazonka.AppSync.DeleteDataSource

Description

Deletes a DataSource object.

Creating a Request

data DeleteDataSource

See: newDeleteDataSource smart constructor.

Constructors

DeleteDataSource'

Fields

apiId :: Text

The API ID.

name :: Text

The name of the data source.

newDeleteDataSource

Arguments

:: Text
$sel:apiId:DeleteDataSource'

-> Text
$sel:name:DeleteDataSource'

-> DeleteDataSource

Create a value of DeleteDataSource with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:apiId:DeleteDataSource', deleteDataSource_apiId - The API ID.

$sel:name:DeleteDataSource', deleteDataSource_name - The name of the data source.

Request Lenses

deleteDataSource_apiId :: Lens' DeleteDataSource Text

The API ID.

deleteDataSource_name :: Lens' DeleteDataSource Text

The name of the data source.

Destructuring the Response

data DeleteDataSourceResponse

See: newDeleteDataSourceResponse smart constructor.

Constructors

DeleteDataSourceResponse'

Fields

httpStatus :: Int

The response's http status code.

newDeleteDataSourceResponse

Arguments

:: Int
$sel:httpStatus:DeleteDataSourceResponse'

-> DeleteDataSourceResponse

Create a value of DeleteDataSourceResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:httpStatus:DeleteDataSourceResponse', deleteDataSourceResponse_httpStatus - The response's http status code.

Response Lenses

deleteDataSourceResponse_httpStatus :: Lens' DeleteDataSourceResponse Int

The response's http status code.