Amazonka.AppSync.GetResolver

Description

Retrieves a Resolver object.

Creating a Request

data GetResolver

See: newGetResolver smart constructor.

Constructors

GetResolver'

Fields

apiId :: Text

The API ID.

typeName :: Text

The resolver type name.

fieldName :: Text

The resolver field name.

newGetResolver

Arguments

:: Text
$sel:apiId:GetResolver'

-> Text
$sel:typeName:GetResolver'

-> Text
$sel:fieldName:GetResolver'

-> GetResolver

Create a value of GetResolver with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:apiId:GetResolver', getResolver_apiId - The API ID.

$sel:typeName:GetResolver', getResolver_typeName - The resolver type name.

$sel:fieldName:GetResolver', getResolver_fieldName - The resolver field name.

Request Lenses

getResolver_apiId :: Lens' GetResolver Text

The API ID.

getResolver_typeName :: Lens' GetResolver Text

The resolver type name.

getResolver_fieldName :: Lens' GetResolver Text

The resolver field name.

Destructuring the Response

data GetResolverResponse

See: newGetResolverResponse smart constructor.

Constructors

GetResolverResponse'

Fields

resolver :: Maybe Resolver

The Resolver object.

httpStatus :: Int

The response's http status code.

newGetResolverResponse

Arguments

:: Int
$sel:httpStatus:GetResolverResponse'

-> GetResolverResponse

Create a value of GetResolverResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:resolver:GetResolverResponse', getResolverResponse_resolver - The Resolver object.

$sel:httpStatus:GetResolverResponse', getResolverResponse_httpStatus - The response's http status code.

Response Lenses

getResolverResponse_resolver :: Lens' GetResolverResponse (Maybe Resolver)

The Resolver object.

getResolverResponse_httpStatus :: Lens' GetResolverResponse Int

The response's http status code.