Amazonka.AppSync.Types.FunctionConfiguration

Description

Documentation

data FunctionConfiguration

A function is a reusable entity. Multiple functions can be used to compose the resolver logic.

See: newFunctionConfiguration smart constructor.

Constructors

FunctionConfiguration'

Fields

functionArn :: Maybe Text

The ARN of the Function object.

dataSourceName :: Maybe Text

The name of the DataSource.

requestMappingTemplate :: Maybe Text

The Function request mapping template. Functions support only the 2018-05-29 version of the request mapping template.

name :: Maybe Text

The name of the Function object.

functionId :: Maybe Text

A unique ID representing the Function object.

responseMappingTemplate :: Maybe Text

The Function response mapping template.

syncConfig :: Maybe SyncConfig

functionVersion :: Maybe Text

The version of the request mapping template. Currently only the 2018-05-29 version of the template is supported.

description :: Maybe Text

The Function description.

newFunctionConfiguration :: FunctionConfiguration

Create a value of FunctionConfiguration with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:functionArn:FunctionConfiguration', functionConfiguration_functionArn - The ARN of the Function object.

$sel:dataSourceName:FunctionConfiguration', functionConfiguration_dataSourceName - The name of the DataSource.

$sel:requestMappingTemplate:FunctionConfiguration', functionConfiguration_requestMappingTemplate - The Function request mapping template. Functions support only the 2018-05-29 version of the request mapping template.

$sel:name:FunctionConfiguration', functionConfiguration_name - The name of the Function object.

$sel:functionId:FunctionConfiguration', functionConfiguration_functionId - A unique ID representing the Function object.

$sel:responseMappingTemplate:FunctionConfiguration', functionConfiguration_responseMappingTemplate - The Function response mapping template.

$sel:syncConfig:FunctionConfiguration', functionConfiguration_syncConfig - Undocumented member.

$sel:functionVersion:FunctionConfiguration', functionConfiguration_functionVersion - The version of the request mapping template. Currently only the 2018-05-29 version of the template is supported.

$sel:description:FunctionConfiguration', functionConfiguration_description - The Function description.

functionConfiguration_functionArn :: Lens' FunctionConfiguration (Maybe Text)

The ARN of the Function object.

functionConfiguration_dataSourceName :: Lens' FunctionConfiguration (Maybe Text)

The name of the DataSource.

functionConfiguration_requestMappingTemplate :: Lens' FunctionConfiguration (Maybe Text)

The Function request mapping template. Functions support only the 2018-05-29 version of the request mapping template.

functionConfiguration_name :: Lens' FunctionConfiguration (Maybe Text)

The name of the Function object.

functionConfiguration_functionId :: Lens' FunctionConfiguration (Maybe Text)

A unique ID representing the Function object.

functionConfiguration_responseMappingTemplate :: Lens' FunctionConfiguration (Maybe Text)

The Function response mapping template.

functionConfiguration_syncConfig :: Lens' FunctionConfiguration (Maybe SyncConfig)

Undocumented member.

functionConfiguration_functionVersion :: Lens' FunctionConfiguration (Maybe Text)

The version of the request mapping template. Currently only the 2018-05-29 version of the template is supported.

functionConfiguration_description :: Lens' FunctionConfiguration (Maybe Text)

The Function description.