Amazonka.AppSync.Types.Resolver

Description

Documentation

data Resolver

Describes a resolver.

See: newResolver smart constructor.

Constructors

Resolver'

Fields

typeName :: Maybe Text

The resolver type name.

dataSourceName :: Maybe Text

The resolver data source name.

requestMappingTemplate :: Maybe Text

The request mapping template.

kind :: Maybe ResolverKind

The resolver type.

UNIT: A UNIT resolver type. A UNIT resolver is the default resolver type. A UNIT resolver enables you to execute a GraphQL query against a single data source.

PIPELINE: A PIPELINE resolver type. A PIPELINE resolver enables you to execute a series of Function in a serial manner. You can use a pipeline resolver to execute a GraphQL query against multiple data sources.

resolverArn :: Maybe Text

The resolver ARN.

cachingConfig :: Maybe CachingConfig

The caching configuration for the resolver.

responseMappingTemplate :: Maybe Text

The response mapping template.

fieldName :: Maybe Text

The resolver field name.

syncConfig :: Maybe SyncConfig

The SyncConfig for a resolver attached to a versioned datasource.

pipelineConfig :: Maybe PipelineConfig

The PipelineConfig.

newResolver :: Resolver

Create a value of Resolver with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:typeName:Resolver', resolver_typeName - The resolver type name.

$sel:dataSourceName:Resolver', resolver_dataSourceName - The resolver data source name.

$sel:requestMappingTemplate:Resolver', resolver_requestMappingTemplate - The request mapping template.

$sel:kind:Resolver', resolver_kind - The resolver type.

UNIT: A UNIT resolver type. A UNIT resolver is the default resolver type. A UNIT resolver enables you to execute a GraphQL query against a single data source.

PIPELINE: A PIPELINE resolver type. A PIPELINE resolver enables you to execute a series of Function in a serial manner. You can use a pipeline resolver to execute a GraphQL query against multiple data sources.

$sel:resolverArn:Resolver', resolver_resolverArn - The resolver ARN.

$sel:cachingConfig:Resolver', resolver_cachingConfig - The caching configuration for the resolver.

$sel:responseMappingTemplate:Resolver', resolver_responseMappingTemplate - The response mapping template.

$sel:fieldName:Resolver', resolver_fieldName - The resolver field name.

$sel:syncConfig:Resolver', resolver_syncConfig - The SyncConfig for a resolver attached to a versioned datasource.

$sel:pipelineConfig:Resolver', resolver_pipelineConfig - The PipelineConfig.

resolver_typeName :: Lens' Resolver (Maybe Text)

The resolver type name.

resolver_dataSourceName :: Lens' Resolver (Maybe Text)

The resolver data source name.

resolver_requestMappingTemplate :: Lens' Resolver (Maybe Text)

The request mapping template.

resolver_kind :: Lens' Resolver (Maybe ResolverKind)

The resolver type.

UNIT: A UNIT resolver type. A UNIT resolver is the default resolver type. A UNIT resolver enables you to execute a GraphQL query against a single data source.

PIPELINE: A PIPELINE resolver type. A PIPELINE resolver enables you to execute a series of Function in a serial manner. You can use a pipeline resolver to execute a GraphQL query against multiple data sources.

resolver_resolverArn :: Lens' Resolver (Maybe Text)

The resolver ARN.

resolver_cachingConfig :: Lens' Resolver (Maybe CachingConfig)

The caching configuration for the resolver.

resolver_responseMappingTemplate :: Lens' Resolver (Maybe Text)

The response mapping template.

resolver_fieldName :: Lens' Resolver (Maybe Text)

The resolver field name.

resolver_syncConfig :: Lens' Resolver (Maybe SyncConfig)

The SyncConfig for a resolver attached to a versioned datasource.

resolver_pipelineConfig :: Lens' Resolver (Maybe PipelineConfig)

The PipelineConfig.