Amazonka.AppSync.UpdateType

Description

Updates a Type object.

Creating a Request

data UpdateType

See: newUpdateType smart constructor.

Constructors

UpdateType'

Fields

definition :: Maybe Text

The new definition.

apiId :: Text

The API ID.

typeName :: Text

The new type name.

format :: TypeDefinitionFormat

The new type format: SDL or JSON.

newUpdateType

Arguments

:: Text
$sel:apiId:UpdateType'

-> Text
$sel:typeName:UpdateType'

-> TypeDefinitionFormat
$sel:format:UpdateType'

-> UpdateType

Create a value of UpdateType with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:definition:UpdateType', updateType_definition - The new definition.

$sel:apiId:UpdateType', updateType_apiId - The API ID.

$sel:typeName:UpdateType', updateType_typeName - The new type name.

$sel:format:UpdateType', updateType_format - The new type format: SDL or JSON.

Request Lenses

updateType_definition :: Lens' UpdateType (Maybe Text)

The new definition.

updateType_apiId :: Lens' UpdateType Text

The API ID.

updateType_typeName :: Lens' UpdateType Text

The new type name.

updateType_format :: Lens' UpdateType TypeDefinitionFormat

The new type format: SDL or JSON.

Destructuring the Response

data UpdateTypeResponse

See: newUpdateTypeResponse smart constructor.

Constructors

UpdateTypeResponse'

Fields

type' :: Maybe Type

The updated Type object.

httpStatus :: Int

The response's http status code.

newUpdateTypeResponse

Arguments

:: Int
$sel:httpStatus:UpdateTypeResponse'

-> UpdateTypeResponse

Create a value of UpdateTypeResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:type':UpdateTypeResponse', updateTypeResponse_type - The updated Type object.

$sel:httpStatus:UpdateTypeResponse', updateTypeResponse_httpStatus - The response's http status code.

Response Lenses

updateTypeResponse_type :: Lens' UpdateTypeResponse (Maybe Type)

The updated Type object.

updateTypeResponse_httpStatus :: Lens' UpdateTypeResponse Int

The response's http status code.