Amazonka.ApplicationCostProfiler

Description

Derived from API version 2020-09-10 of the AWS service descriptions, licensed under Apache 2.0.

This reference provides descriptions of the AWS Application Cost Profiler API.

The AWS Application Cost Profiler API provides programmatic access to view, create, update, and delete application cost report definitions, as well as to import your usage data into the Application Cost Profiler service.

For more information about using this service, see the AWS Application Cost Profiler User Guide.

Service Configuration

defaultService :: Service

API version 2020-09-10 of the Amazon Application Cost Profiler SDK configuration.

Errors

Error matchers are designed for use with the functions provided by Control.Exception.Lens. This allows catching (and rethrowing) service specific errors returned by ApplicationCostProfiler.

ValidationException

_ValidationException :: AsError a => Getting (First ServiceError) a ServiceError

The input fails to satisfy the constraints for the API.

AccessDeniedException

_AccessDeniedException :: AsError a => Getting (First ServiceError) a ServiceError

You do not have permission to perform this action.

ServiceQuotaExceededException

_ServiceQuotaExceededException :: AsError a => Getting (First ServiceError) a ServiceError

Your request exceeds one or more of the service quotas.

ThrottlingException

_ThrottlingException :: AsError a => Getting (First ServiceError) a ServiceError

The calls to AWS Application Cost Profiler API are throttled. The request was denied.

InternalServerException

_InternalServerException :: AsError a => Getting (First ServiceError) a ServiceError

An internal server error occurred. Retry your request.

Waiters

Waiters poll by repeatedly sending a request until some remote success condition configured by the Wait specification is fulfilled. The Wait specification determines how many attempts should be made, in addition to delay and retry strategies.

Operations

Some AWS operations return results that are incomplete and require subsequent requests in order to obtain the entire result set. The process of sending subsequent requests to continue where a previous request left off is called pagination. For example, the ListObjects operation of Amazon S3 returns up to 1000 objects at a time, and you must send subsequent requests with the appropriate Marker in order to retrieve the next page of results.

Operations that have an AWSPager instance can transparently perform subsequent requests, correctly setting Markers and other request facets to iterate through the entire result set of a truncated API operation. Operations which support this have an additional note in the documentation.

Many operations have the ability to filter results on the server side. See the individual operation parameters for details.

ImportApplicationUsage

data ImportApplicationUsage

See: newImportApplicationUsage smart constructor.

Constructors

ImportApplicationUsage' SourceS3Location

newImportApplicationUsage

Arguments

:: SourceS3Location
$sel:sourceS3Location:ImportApplicationUsage'

-> ImportApplicationUsage

Create a value of ImportApplicationUsage with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:sourceS3Location:ImportApplicationUsage', importApplicationUsage_sourceS3Location - Amazon S3 location to import application usage data from.

data ImportApplicationUsageResponse

See: newImportApplicationUsageResponse smart constructor.

Constructors

ImportApplicationUsageResponse' Int Text

newImportApplicationUsageResponse

Arguments

:: Int
$sel:httpStatus:ImportApplicationUsageResponse'

-> Text
$sel:importId:ImportApplicationUsageResponse'

-> ImportApplicationUsageResponse

Create a value of ImportApplicationUsageResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:httpStatus:ImportApplicationUsageResponse', importApplicationUsageResponse_httpStatus - The response's http status code.

$sel:importId:ImportApplicationUsageResponse', importApplicationUsageResponse_importId - ID of the import request.

PutReportDefinition

data PutReportDefinition

See: newPutReportDefinition smart constructor.

Constructors

PutReportDefinition' Text Text ReportFrequency Format S3Location

newPutReportDefinition

Arguments

:: Text
$sel:reportId:PutReportDefinition'

-> Text
$sel:reportDescription:PutReportDefinition'

-> ReportFrequency
$sel:reportFrequency:PutReportDefinition'

-> Format
$sel:format:PutReportDefinition'

-> S3Location
$sel:destinationS3Location:PutReportDefinition'

-> PutReportDefinition

Create a value of PutReportDefinition with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:reportId:PutReportDefinition', putReportDefinition_reportId - Required. ID of the report. You can choose any valid string matching the pattern for the ID.

$sel:reportDescription:PutReportDefinition', putReportDefinition_reportDescription - Required. Description of the report.

$sel:reportFrequency:PutReportDefinition', putReportDefinition_reportFrequency - Required. The cadence to generate the report.

$sel:format:PutReportDefinition', putReportDefinition_format - Required. The format to use for the generated report.

$sel:destinationS3Location:PutReportDefinition', putReportDefinition_destinationS3Location - Required. Amazon Simple Storage Service (Amazon S3) location where Application Cost Profiler uploads the report.

data PutReportDefinitionResponse

See: newPutReportDefinitionResponse smart constructor.

Constructors

PutReportDefinitionResponse' (Maybe Text) Int

newPutReportDefinitionResponse

Arguments

:: Int
$sel:httpStatus:PutReportDefinitionResponse'

-> PutReportDefinitionResponse

Create a value of PutReportDefinitionResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:reportId:PutReportDefinition', putReportDefinitionResponse_reportId - ID of the report.

$sel:httpStatus:PutReportDefinitionResponse', putReportDefinitionResponse_httpStatus - The response's http status code.

DeleteReportDefinition

data DeleteReportDefinition

See: newDeleteReportDefinition smart constructor.

Constructors

DeleteReportDefinition' Text

newDeleteReportDefinition

Arguments

:: Text
$sel:reportId:DeleteReportDefinition'

-> DeleteReportDefinition

Create a value of DeleteReportDefinition with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:reportId:DeleteReportDefinition', deleteReportDefinition_reportId - Required. ID of the report to delete.

data DeleteReportDefinitionResponse

See: newDeleteReportDefinitionResponse smart constructor.

Constructors

DeleteReportDefinitionResponse' (Maybe Text) Int

newDeleteReportDefinitionResponse

Arguments

:: Int
$sel:httpStatus:DeleteReportDefinitionResponse'

-> DeleteReportDefinitionResponse

Create a value of DeleteReportDefinitionResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:reportId:DeleteReportDefinition', deleteReportDefinitionResponse_reportId - ID of the report that was deleted.

$sel:httpStatus:DeleteReportDefinitionResponse', deleteReportDefinitionResponse_httpStatus - The response's http status code.

UpdateReportDefinition

data UpdateReportDefinition

See: newUpdateReportDefinition smart constructor.

Constructors

UpdateReportDefinition' Text Text ReportFrequency Format S3Location

newUpdateReportDefinition

Arguments

:: Text
$sel:reportId:UpdateReportDefinition'

-> Text
$sel:reportDescription:UpdateReportDefinition'

-> ReportFrequency
$sel:reportFrequency:UpdateReportDefinition'

-> Format
$sel:format:UpdateReportDefinition'

-> S3Location
$sel:destinationS3Location:UpdateReportDefinition'

-> UpdateReportDefinition

Create a value of UpdateReportDefinition with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:reportId:UpdateReportDefinition', updateReportDefinition_reportId - Required. ID of the report to update.

$sel:reportDescription:UpdateReportDefinition', updateReportDefinition_reportDescription - Required. Description of the report.

$sel:reportFrequency:UpdateReportDefinition', updateReportDefinition_reportFrequency - Required. The cadence to generate the report.

$sel:format:UpdateReportDefinition', updateReportDefinition_format - Required. The format to use for the generated report.

$sel:destinationS3Location:UpdateReportDefinition', updateReportDefinition_destinationS3Location - Required. Amazon Simple Storage Service (Amazon S3) location where Application Cost Profiler uploads the report.

data UpdateReportDefinitionResponse

See: newUpdateReportDefinitionResponse smart constructor.

Constructors

UpdateReportDefinitionResponse' (Maybe Text) Int

newUpdateReportDefinitionResponse

Arguments

:: Int
$sel:httpStatus:UpdateReportDefinitionResponse'

-> UpdateReportDefinitionResponse

Create a value of UpdateReportDefinitionResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:reportId:UpdateReportDefinition', updateReportDefinitionResponse_reportId - ID of the report.

$sel:httpStatus:UpdateReportDefinitionResponse', updateReportDefinitionResponse_httpStatus - The response's http status code.

GetReportDefinition

data GetReportDefinition

See: newGetReportDefinition smart constructor.

Constructors

GetReportDefinition' Text

newGetReportDefinition

Arguments

:: Text
$sel:reportId:GetReportDefinition'

-> GetReportDefinition

Create a value of GetReportDefinition with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:reportId:GetReportDefinition', getReportDefinition_reportId - ID of the report to retrieve.

data GetReportDefinitionResponse

See: newGetReportDefinitionResponse smart constructor.

Constructors

GetReportDefinitionResponse' Int Text Text ReportFrequency Format S3Location POSIX POSIX

newGetReportDefinitionResponse

Arguments

:: Int
$sel:httpStatus:GetReportDefinitionResponse'

-> Text
$sel:reportId:GetReportDefinition'

-> Text
$sel:reportDescription:GetReportDefinitionResponse'

-> ReportFrequency
$sel:reportFrequency:GetReportDefinitionResponse'

-> Format
$sel:format:GetReportDefinitionResponse'

-> S3Location
$sel:destinationS3Location:GetReportDefinitionResponse'

-> UTCTime
$sel:createdAt:GetReportDefinitionResponse'

-> UTCTime
$sel:lastUpdated:GetReportDefinitionResponse'

-> GetReportDefinitionResponse

Create a value of GetReportDefinitionResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:httpStatus:GetReportDefinitionResponse', getReportDefinitionResponse_httpStatus - The response's http status code.

$sel:reportId:GetReportDefinition', getReportDefinitionResponse_reportId - ID of the report retrieved.

$sel:reportDescription:GetReportDefinitionResponse', getReportDefinitionResponse_reportDescription - Description of the report.

$sel:reportFrequency:GetReportDefinitionResponse', getReportDefinitionResponse_reportFrequency - Cadence used to generate the report.

$sel:format:GetReportDefinitionResponse', getReportDefinitionResponse_format - Format of the generated report.

$sel:destinationS3Location:GetReportDefinitionResponse', getReportDefinitionResponse_destinationS3Location - Amazon Simple Storage Service (Amazon S3) location where the report is uploaded.

$sel:createdAt:GetReportDefinitionResponse', getReportDefinitionResponse_createdAt - Timestamp (milliseconds) when this report definition was created.

$sel:lastUpdated:GetReportDefinitionResponse', getReportDefinitionResponse_lastUpdated - Timestamp (milliseconds) when this report definition was last updated.

ListReportDefinitions (Paginated)

data ListReportDefinitions

See: newListReportDefinitions smart constructor.

Constructors

ListReportDefinitions' (Maybe Text) (Maybe Natural)

newListReportDefinitions :: ListReportDefinitions

Create a value of ListReportDefinitions with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:nextToken:ListReportDefinitions', listReportDefinitions_nextToken - The token value from a previous call to access the next page of results.

$sel:maxResults:ListReportDefinitions', listReportDefinitions_maxResults - The maximum number of results to return.

data ListReportDefinitionsResponse

See: newListReportDefinitionsResponse smart constructor.

Constructors

ListReportDefinitionsResponse' (Maybe Text) (Maybe [ReportDefinition]) Int

newListReportDefinitionsResponse

Arguments

:: Int
$sel:httpStatus:ListReportDefinitionsResponse'

-> ListReportDefinitionsResponse

Create a value of ListReportDefinitionsResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:nextToken:ListReportDefinitions', listReportDefinitionsResponse_nextToken - The value of the next token, if it exists. Null if there are no more results.

$sel:reportDefinitions:ListReportDefinitionsResponse', listReportDefinitionsResponse_reportDefinitions - The retrieved reports.

$sel:httpStatus:ListReportDefinitionsResponse', listReportDefinitionsResponse_httpStatus - The response's http status code.

Types

Format

newtype Format

Constructors

Format'

Fields

fromFormat :: Text

Bundled Patterns

pattern Format_CSV :: Format

pattern Format_PARQUET :: Format

ReportFrequency

newtype ReportFrequency

Constructors

ReportFrequency'

Fields

fromReportFrequency :: Text

Bundled Patterns

pattern ReportFrequency_ALL :: ReportFrequency

pattern ReportFrequency_DAILY :: ReportFrequency

pattern ReportFrequency_MONTHLY :: ReportFrequency

S3BucketRegion

newtype S3BucketRegion

Constructors

S3BucketRegion'

Fields

fromS3BucketRegion :: Text

Bundled Patterns

pattern S3BucketRegion_Af_south_1 :: S3BucketRegion

pattern S3BucketRegion_Ap_east_1 :: S3BucketRegion

pattern S3BucketRegion_Eu_south_1 :: S3BucketRegion

pattern S3BucketRegion_Me_south_1 :: S3BucketRegion

ReportDefinition

data ReportDefinition

The configuration of a report in AWS Application Cost Profiler.

See: newReportDefinition smart constructor.

Constructors

ReportDefinition' (Maybe POSIX) (Maybe POSIX) (Maybe Text) (Maybe Format) (Maybe ReportFrequency) (Maybe Text) (Maybe S3Location)

newReportDefinition :: ReportDefinition

Create a value of ReportDefinition with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:lastUpdatedAt:ReportDefinition', reportDefinition_lastUpdatedAt - Timestamp (milliseconds) when this report definition was last updated.

$sel:createdAt:ReportDefinition', reportDefinition_createdAt - Timestamp (milliseconds) when this report definition was created.

$sel:reportId:ReportDefinition', reportDefinition_reportId - The ID of the report.

$sel:format:ReportDefinition', reportDefinition_format - The format used for the generated reports.

$sel:reportFrequency:ReportDefinition', reportDefinition_reportFrequency - The cadence at which the report is generated.

$sel:reportDescription:ReportDefinition', reportDefinition_reportDescription - Description of the report

$sel:destinationS3Location:ReportDefinition', reportDefinition_destinationS3Location - The location in Amazon Simple Storage Service (Amazon S3) the reports should be saved to.

S3Location

data S3Location

Represents the Amazon Simple Storage Service (Amazon S3) location where AWS Application Cost Profiler reports are generated and then written to.

See: newS3Location smart constructor.

Constructors

S3Location' Text Text

newS3Location

Arguments

:: Text
$sel:bucket:S3Location'

-> Text
$sel:prefix:S3Location'

-> S3Location

Create a value of S3Location with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:bucket:S3Location', s3Location_bucket - Name of the S3 bucket.

$sel:prefix:S3Location', s3Location_prefix - Prefix for the location to write to.

SourceS3Location

data SourceS3Location

Represents the Amazon Simple Storage Service (Amazon S3) location where usage data is read from.

See: newSourceS3Location smart constructor.

Constructors

SourceS3Location' (Maybe S3BucketRegion) Text Text

newSourceS3Location

Arguments

:: Text
$sel:bucket:SourceS3Location'

-> Text
$sel:key:SourceS3Location'

-> SourceS3Location

Create a value of SourceS3Location with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:region:SourceS3Location', sourceS3Location_region - Region of the bucket. Only required for Regions that are disabled by default. For more infomration about Regions that are disabled by default, see Enabling a Region in the AWS General Reference guide.

$sel:bucket:SourceS3Location', sourceS3Location_bucket - Name of the bucket.

$sel:key:SourceS3Location', sourceS3Location_key - Key of the object.