Amazonka.ApplicationCostProfiler.Lens

Description

Operations

ImportApplicationUsage

importApplicationUsage_sourceS3Location :: Lens' ImportApplicationUsage SourceS3Location

Amazon S3 location to import application usage data from.

importApplicationUsageResponse_httpStatus :: Lens' ImportApplicationUsageResponse Int

The response's http status code.

importApplicationUsageResponse_importId :: Lens' ImportApplicationUsageResponse Text

ID of the import request.

PutReportDefinition

putReportDefinition_reportId :: Lens' PutReportDefinition Text

Required. ID of the report. You can choose any valid string matching the pattern for the ID.

putReportDefinition_reportDescription :: Lens' PutReportDefinition Text

Required. Description of the report.

putReportDefinition_reportFrequency :: Lens' PutReportDefinition ReportFrequency

Required. The cadence to generate the report.

putReportDefinition_format :: Lens' PutReportDefinition Format

Required. The format to use for the generated report.

putReportDefinition_destinationS3Location :: Lens' PutReportDefinition S3Location

Required. Amazon Simple Storage Service (Amazon S3) location where Application Cost Profiler uploads the report.

putReportDefinitionResponse_reportId :: Lens' PutReportDefinitionResponse (Maybe Text)

ID of the report.

putReportDefinitionResponse_httpStatus :: Lens' PutReportDefinitionResponse Int

The response's http status code.

DeleteReportDefinition

deleteReportDefinition_reportId :: Lens' DeleteReportDefinition Text

Required. ID of the report to delete.

deleteReportDefinitionResponse_reportId :: Lens' DeleteReportDefinitionResponse (Maybe Text)

ID of the report that was deleted.

deleteReportDefinitionResponse_httpStatus :: Lens' DeleteReportDefinitionResponse Int

The response's http status code.

UpdateReportDefinition

updateReportDefinition_reportId :: Lens' UpdateReportDefinition Text

Required. ID of the report to update.

updateReportDefinition_reportDescription :: Lens' UpdateReportDefinition Text

Required. Description of the report.

updateReportDefinition_reportFrequency :: Lens' UpdateReportDefinition ReportFrequency

Required. The cadence to generate the report.

updateReportDefinition_format :: Lens' UpdateReportDefinition Format

Required. The format to use for the generated report.

updateReportDefinition_destinationS3Location :: Lens' UpdateReportDefinition S3Location

Required. Amazon Simple Storage Service (Amazon S3) location where Application Cost Profiler uploads the report.

updateReportDefinitionResponse_reportId :: Lens' UpdateReportDefinitionResponse (Maybe Text)

ID of the report.

updateReportDefinitionResponse_httpStatus :: Lens' UpdateReportDefinitionResponse Int

The response's http status code.

GetReportDefinition

getReportDefinition_reportId :: Lens' GetReportDefinition Text

ID of the report to retrieve.

getReportDefinitionResponse_httpStatus :: Lens' GetReportDefinitionResponse Int

The response's http status code.

getReportDefinitionResponse_reportId :: Lens' GetReportDefinitionResponse Text

ID of the report retrieved.

getReportDefinitionResponse_reportDescription :: Lens' GetReportDefinitionResponse Text

Description of the report.

getReportDefinitionResponse_reportFrequency :: Lens' GetReportDefinitionResponse ReportFrequency

Cadence used to generate the report.

getReportDefinitionResponse_format :: Lens' GetReportDefinitionResponse Format

Format of the generated report.

getReportDefinitionResponse_destinationS3Location :: Lens' GetReportDefinitionResponse S3Location

Amazon Simple Storage Service (Amazon S3) location where the report is uploaded.

getReportDefinitionResponse_createdAt :: Lens' GetReportDefinitionResponse UTCTime

Timestamp (milliseconds) when this report definition was created.

getReportDefinitionResponse_lastUpdated :: Lens' GetReportDefinitionResponse UTCTime

Timestamp (milliseconds) when this report definition was last updated.

ListReportDefinitions

listReportDefinitions_nextToken :: Lens' ListReportDefinitions (Maybe Text)

The token value from a previous call to access the next page of results.

listReportDefinitions_maxResults :: Lens' ListReportDefinitions (Maybe Natural)

The maximum number of results to return.

listReportDefinitionsResponse_nextToken :: Lens' ListReportDefinitionsResponse (Maybe Text)

The value of the next token, if it exists. Null if there are no more results.

listReportDefinitionsResponse_reportDefinitions :: Lens' ListReportDefinitionsResponse (Maybe [ReportDefinition])

The retrieved reports.

listReportDefinitionsResponse_httpStatus :: Lens' ListReportDefinitionsResponse Int

The response's http status code.

Types

ReportDefinition

reportDefinition_lastUpdatedAt :: Lens' ReportDefinition (Maybe UTCTime)

Timestamp (milliseconds) when this report definition was last updated.

reportDefinition_createdAt :: Lens' ReportDefinition (Maybe UTCTime)

Timestamp (milliseconds) when this report definition was created.

reportDefinition_reportId :: Lens' ReportDefinition (Maybe Text)

The ID of the report.

reportDefinition_format :: Lens' ReportDefinition (Maybe Format)

The format used for the generated reports.

reportDefinition_reportFrequency :: Lens' ReportDefinition (Maybe ReportFrequency)

The cadence at which the report is generated.

reportDefinition_reportDescription :: Lens' ReportDefinition (Maybe Text)

Description of the report

reportDefinition_destinationS3Location :: Lens' ReportDefinition (Maybe S3Location)

The location in Amazon Simple Storage Service (Amazon S3) the reports should be saved to.

S3Location

s3Location_bucket :: Lens' S3Location Text

Name of the S3 bucket.

s3Location_prefix :: Lens' S3Location Text

Prefix for the location to write to.

SourceS3Location

sourceS3Location_region :: Lens' SourceS3Location (Maybe S3BucketRegion)

Region of the bucket. Only required for Regions that are disabled by default. For more infomration about Regions that are disabled by default, see Enabling a Region in the AWS General Reference guide.

sourceS3Location_bucket :: Lens' SourceS3Location Text

Name of the bucket.

sourceS3Location_key :: Lens' SourceS3Location Text

Key of the object.