Amazonka.ApplicationCostProfiler.Types

Description

Service Configuration

defaultService :: Service

API version 2020-09-10 of the Amazon Application Cost Profiler SDK configuration.

Errors

_ValidationException :: AsError a => Getting (First ServiceError) a ServiceError

The input fails to satisfy the constraints for the API.

_AccessDeniedException :: AsError a => Getting (First ServiceError) a ServiceError

You do not have permission to perform this action.

_ServiceQuotaExceededException :: AsError a => Getting (First ServiceError) a ServiceError

Your request exceeds one or more of the service quotas.

_ThrottlingException :: AsError a => Getting (First ServiceError) a ServiceError

The calls to AWS Application Cost Profiler API are throttled. The request was denied.

_InternalServerException :: AsError a => Getting (First ServiceError) a ServiceError

An internal server error occurred. Retry your request.

Format

newtype Format

Constructors

Format'

Fields

fromFormat :: Text

Bundled Patterns

pattern Format_CSV :: Format

pattern Format_PARQUET :: Format

ReportFrequency

newtype ReportFrequency

Constructors

ReportFrequency'

Fields

fromReportFrequency :: Text

Bundled Patterns

pattern ReportFrequency_ALL :: ReportFrequency

pattern ReportFrequency_DAILY :: ReportFrequency

pattern ReportFrequency_MONTHLY :: ReportFrequency

S3BucketRegion

newtype S3BucketRegion

Constructors

S3BucketRegion'

Fields

fromS3BucketRegion :: Text

Bundled Patterns

pattern S3BucketRegion_Af_south_1 :: S3BucketRegion

pattern S3BucketRegion_Ap_east_1 :: S3BucketRegion

pattern S3BucketRegion_Eu_south_1 :: S3BucketRegion

pattern S3BucketRegion_Me_south_1 :: S3BucketRegion

ReportDefinition

data ReportDefinition

The configuration of a report in AWS Application Cost Profiler.

See: newReportDefinition smart constructor.

Constructors

ReportDefinition'

Fields

lastUpdatedAt :: Maybe POSIX

Timestamp (milliseconds) when this report definition was last updated.

createdAt :: Maybe POSIX

Timestamp (milliseconds) when this report definition was created.

reportId :: Maybe Text

The ID of the report.

format :: Maybe Format

The format used for the generated reports.

reportFrequency :: Maybe ReportFrequency

The cadence at which the report is generated.

reportDescription :: Maybe Text

Description of the report

destinationS3Location :: Maybe S3Location

The location in Amazon Simple Storage Service (Amazon S3) the reports should be saved to.

newReportDefinition :: ReportDefinition

Create a value of ReportDefinition with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:lastUpdatedAt:ReportDefinition', reportDefinition_lastUpdatedAt - Timestamp (milliseconds) when this report definition was last updated.

$sel:createdAt:ReportDefinition', reportDefinition_createdAt - Timestamp (milliseconds) when this report definition was created.

$sel:reportId:ReportDefinition', reportDefinition_reportId - The ID of the report.

$sel:format:ReportDefinition', reportDefinition_format - The format used for the generated reports.

$sel:reportFrequency:ReportDefinition', reportDefinition_reportFrequency - The cadence at which the report is generated.

$sel:reportDescription:ReportDefinition', reportDefinition_reportDescription - Description of the report

$sel:destinationS3Location:ReportDefinition', reportDefinition_destinationS3Location - The location in Amazon Simple Storage Service (Amazon S3) the reports should be saved to.

reportDefinition_lastUpdatedAt :: Lens' ReportDefinition (Maybe UTCTime)

Timestamp (milliseconds) when this report definition was last updated.

reportDefinition_createdAt :: Lens' ReportDefinition (Maybe UTCTime)

Timestamp (milliseconds) when this report definition was created.

reportDefinition_reportId :: Lens' ReportDefinition (Maybe Text)

The ID of the report.

reportDefinition_format :: Lens' ReportDefinition (Maybe Format)

The format used for the generated reports.

reportDefinition_reportFrequency :: Lens' ReportDefinition (Maybe ReportFrequency)

The cadence at which the report is generated.

reportDefinition_reportDescription :: Lens' ReportDefinition (Maybe Text)

Description of the report

reportDefinition_destinationS3Location :: Lens' ReportDefinition (Maybe S3Location)

The location in Amazon Simple Storage Service (Amazon S3) the reports should be saved to.

S3Location

data S3Location

Represents the Amazon Simple Storage Service (Amazon S3) location where AWS Application Cost Profiler reports are generated and then written to.

See: newS3Location smart constructor.

Constructors

S3Location'

Fields

bucket :: Text

Name of the S3 bucket.

prefix :: Text

Prefix for the location to write to.

newS3Location

Arguments

:: Text
$sel:bucket:S3Location'

-> Text
$sel:prefix:S3Location'

-> S3Location

Create a value of S3Location with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:bucket:S3Location', s3Location_bucket - Name of the S3 bucket.

$sel:prefix:S3Location', s3Location_prefix - Prefix for the location to write to.

s3Location_bucket :: Lens' S3Location Text

Name of the S3 bucket.

s3Location_prefix :: Lens' S3Location Text

Prefix for the location to write to.

SourceS3Location

data SourceS3Location

Represents the Amazon Simple Storage Service (Amazon S3) location where usage data is read from.

See: newSourceS3Location smart constructor.

Constructors

SourceS3Location'

Fields

region :: Maybe S3BucketRegion

Region of the bucket. Only required for Regions that are disabled by default. For more infomration about Regions that are disabled by default, see Enabling a Region in the AWS General Reference guide.

bucket :: Text

Name of the bucket.

key :: Text

Key of the object.

newSourceS3Location

Arguments

:: Text
$sel:bucket:SourceS3Location'

-> Text
$sel:key:SourceS3Location'

-> SourceS3Location

Create a value of SourceS3Location with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:region:SourceS3Location', sourceS3Location_region - Region of the bucket. Only required for Regions that are disabled by default. For more infomration about Regions that are disabled by default, see Enabling a Region in the AWS General Reference guide.

$sel:bucket:SourceS3Location', sourceS3Location_bucket - Name of the bucket.

$sel:key:SourceS3Location', sourceS3Location_key - Key of the object.

sourceS3Location_region :: Lens' SourceS3Location (Maybe S3BucketRegion)

Region of the bucket. Only required for Regions that are disabled by default. For more infomration about Regions that are disabled by default, see Enabling a Region in the AWS General Reference guide.

sourceS3Location_bucket :: Lens' SourceS3Location Text

Name of the bucket.

sourceS3Location_key :: Lens' SourceS3Location Text

Key of the object.