Amazonka.ApplicationInsights.CreateApplication

Description

Adds an application that is created from a resource group.

Creating a Request

data CreateApplication

See: newCreateApplication smart constructor.

Constructors

CreateApplication'

Fields

cWEMonitorEnabled :: Maybe Bool

Indicates whether Application Insights can listen to CloudWatch events for the application resources, such as instance terminated, failed deployment, and others.

opsItemSNSTopicArn :: Maybe Text

The SNS topic provided to Application Insights that is associated to the created opsItem. Allows you to receive notifications for updates to the opsItem.

opsCenterEnabled :: Maybe Bool

When set to true, creates opsItems for any problems detected on an application.

tags :: Maybe [Tag]

List of tags to add to the application. tag key (Key) and an associated tag value (Value). The maximum length of a tag key is 128 characters. The maximum length of a tag value is 256 characters.

resourceGroupName :: Text

The name of the resource group.

newCreateApplication

Arguments

:: Text
$sel:resourceGroupName:CreateApplication'

-> CreateApplication

Create a value of CreateApplication with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:cWEMonitorEnabled:CreateApplication', createApplication_cWEMonitorEnabled - Indicates whether Application Insights can listen to CloudWatch events for the application resources, such as instance terminated, failed deployment, and others.

$sel:opsItemSNSTopicArn:CreateApplication', createApplication_opsItemSNSTopicArn - The SNS topic provided to Application Insights that is associated to the created opsItem. Allows you to receive notifications for updates to the opsItem.

$sel:opsCenterEnabled:CreateApplication', createApplication_opsCenterEnabled - When set to true, creates opsItems for any problems detected on an application.

$sel:tags:CreateApplication', createApplication_tags - List of tags to add to the application. tag key (Key) and an associated tag value (Value). The maximum length of a tag key is 128 characters. The maximum length of a tag value is 256 characters.

$sel:resourceGroupName:CreateApplication', createApplication_resourceGroupName - The name of the resource group.

Request Lenses

createApplication_cWEMonitorEnabled :: Lens' CreateApplication (Maybe Bool)

Indicates whether Application Insights can listen to CloudWatch events for the application resources, such as instance terminated, failed deployment, and others.

createApplication_opsItemSNSTopicArn :: Lens' CreateApplication (Maybe Text)

The SNS topic provided to Application Insights that is associated to the created opsItem. Allows you to receive notifications for updates to the opsItem.

createApplication_opsCenterEnabled :: Lens' CreateApplication (Maybe Bool)

When set to true, creates opsItems for any problems detected on an application.

createApplication_tags :: Lens' CreateApplication (Maybe [Tag])

List of tags to add to the application. tag key (Key) and an associated tag value (Value). The maximum length of a tag key is 128 characters. The maximum length of a tag value is 256 characters.

createApplication_resourceGroupName :: Lens' CreateApplication Text

The name of the resource group.

Destructuring the Response

data CreateApplicationResponse

See: newCreateApplicationResponse smart constructor.

Constructors

CreateApplicationResponse'

Fields

applicationInfo :: Maybe ApplicationInfo

Information about the application.

httpStatus :: Int

The response's http status code.

newCreateApplicationResponse

Arguments

:: Int
$sel:httpStatus:CreateApplicationResponse'

-> CreateApplicationResponse

Create a value of CreateApplicationResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:applicationInfo:CreateApplicationResponse', createApplicationResponse_applicationInfo - Information about the application.

$sel:httpStatus:CreateApplicationResponse', createApplicationResponse_httpStatus - The response's http status code.

Response Lenses

createApplicationResponse_applicationInfo :: Lens' CreateApplicationResponse (Maybe ApplicationInfo)

Information about the application.

createApplicationResponse_httpStatus :: Lens' CreateApplicationResponse Int

The response's http status code.