Amazonka.ApplicationInsights.Lens

Description

Operations

DescribeApplication

describeApplication_resourceGroupName :: Lens' DescribeApplication Text

The name of the resource group.

describeApplicationResponse_applicationInfo :: Lens' DescribeApplicationResponse (Maybe ApplicationInfo)

Information about the application.

describeApplicationResponse_httpStatus :: Lens' DescribeApplicationResponse Int

The response's http status code.

DescribeComponent

describeComponent_resourceGroupName :: Lens' DescribeComponent Text

The name of the resource group.

describeComponent_componentName :: Lens' DescribeComponent Text

The name of the component.

describeComponentResponse_applicationComponent :: Lens' DescribeComponentResponse (Maybe ApplicationComponent)

Undocumented member.

describeComponentResponse_resourceList :: Lens' DescribeComponentResponse (Maybe [Text])

The list of resource ARNs that belong to the component.

describeComponentResponse_httpStatus :: Lens' DescribeComponentResponse Int

The response's http status code.

ListTagsForResource

listTagsForResource_resourceARN :: Lens' ListTagsForResource Text

The Amazon Resource Name (ARN) of the application that you want to retrieve tag information for.

listTagsForResourceResponse_tags :: Lens' ListTagsForResourceResponse (Maybe [Tag])

An array that lists all the tags that are associated with the application. Each tag consists of a required tag key (Key) and an associated tag value (Value).

listTagsForResourceResponse_httpStatus :: Lens' ListTagsForResourceResponse Int

The response's http status code.

DeleteApplication

deleteApplication_resourceGroupName :: Lens' DeleteApplication Text

The name of the resource group.

deleteApplicationResponse_httpStatus :: Lens' DeleteApplicationResponse Int

The response's http status code.

UpdateApplication

updateApplication_cWEMonitorEnabled :: Lens' UpdateApplication (Maybe Bool)

Indicates whether Application Insights can listen to CloudWatch events for the application resources, such as instance terminated, failed deployment, and others.

updateApplication_opsItemSNSTopicArn :: Lens' UpdateApplication (Maybe Text)

The SNS topic provided to Application Insights that is associated to the created opsItem. Allows you to receive notifications for updates to the opsItem.

updateApplication_removeSNSTopic :: Lens' UpdateApplication (Maybe Bool)

Disassociates the SNS topic from the opsItem created for detected problems.

updateApplication_opsCenterEnabled :: Lens' UpdateApplication (Maybe Bool)

When set to true, creates opsItems for any problems detected on an application.

updateApplication_resourceGroupName :: Lens' UpdateApplication Text

The name of the resource group.

updateApplicationResponse_applicationInfo :: Lens' UpdateApplicationResponse (Maybe ApplicationInfo)

Information about the application.

updateApplicationResponse_httpStatus :: Lens' UpdateApplicationResponse Int

The response's http status code.

DescribeComponentConfigurationRecommendation

describeComponentConfigurationRecommendation_resourceGroupName :: Lens' DescribeComponentConfigurationRecommendation Text

The name of the resource group.

describeComponentConfigurationRecommendation_componentName :: Lens' DescribeComponentConfigurationRecommendation Text

The name of the component.

describeComponentConfigurationRecommendation_tier :: Lens' DescribeComponentConfigurationRecommendation Tier

The tier of the application component. Supported tiers include DOT_NET_CORE, DOT_NET_WORKER, DOT_NET_WEB, SQL_SERVER, and DEFAULT.

describeComponentConfigurationRecommendationResponse_componentConfiguration :: Lens' DescribeComponentConfigurationRecommendationResponse (Maybe Text)

The recommended configuration settings of the component. The value is the escaped JSON of the configuration.

describeComponentConfigurationRecommendationResponse_httpStatus :: Lens' DescribeComponentConfigurationRecommendationResponse Int

The response's http status code.

DescribeProblem

describeProblem_problemId :: Lens' DescribeProblem Text

The ID of the problem.

describeProblemResponse_problem :: Lens' DescribeProblemResponse (Maybe Problem)

Information about the problem.

describeProblemResponse_httpStatus :: Lens' DescribeProblemResponse Int

The response's http status code.

UpdateComponentConfiguration

updateComponentConfiguration_componentConfiguration :: Lens' UpdateComponentConfiguration (Maybe Text)

The configuration settings of the component. The value is the escaped JSON of the configuration. For more information about the JSON format, see Working with JSON. You can send a request to DescribeComponentConfigurationRecommendation to see the recommended configuration for a component. For the complete format of the component configuration file, see Component Configuration.

updateComponentConfiguration_monitor :: Lens' UpdateComponentConfiguration (Maybe Bool)

Indicates whether the application component is monitored.

updateComponentConfiguration_tier :: Lens' UpdateComponentConfiguration (Maybe Tier)

The tier of the application component. Supported tiers include DOT_NET_WORKER, DOT_NET_WEB, DOT_NET_CORE, SQL_SERVER, and DEFAULT.

updateComponentConfiguration_resourceGroupName :: Lens' UpdateComponentConfiguration Text

The name of the resource group.

updateComponentConfiguration_componentName :: Lens' UpdateComponentConfiguration Text

The name of the component.

updateComponentConfigurationResponse_httpStatus :: Lens' UpdateComponentConfigurationResponse Int

The response's http status code.

CreateApplication

createApplication_cWEMonitorEnabled :: Lens' CreateApplication (Maybe Bool)

Indicates whether Application Insights can listen to CloudWatch events for the application resources, such as instance terminated, failed deployment, and others.

createApplication_opsItemSNSTopicArn :: Lens' CreateApplication (Maybe Text)

The SNS topic provided to Application Insights that is associated to the created opsItem. Allows you to receive notifications for updates to the opsItem.

createApplication_opsCenterEnabled :: Lens' CreateApplication (Maybe Bool)

When set to true, creates opsItems for any problems detected on an application.

createApplication_tags :: Lens' CreateApplication (Maybe [Tag])

List of tags to add to the application. tag key (Key) and an associated tag value (Value). The maximum length of a tag key is 128 characters. The maximum length of a tag value is 256 characters.

createApplication_resourceGroupName :: Lens' CreateApplication Text

The name of the resource group.

createApplicationResponse_applicationInfo :: Lens' CreateApplicationResponse (Maybe ApplicationInfo)

Information about the application.

createApplicationResponse_httpStatus :: Lens' CreateApplicationResponse Int

The response's http status code.

DescribeProblemObservations

describeProblemObservations_problemId :: Lens' DescribeProblemObservations Text

The ID of the problem.

describeProblemObservationsResponse_relatedObservations :: Lens' DescribeProblemObservationsResponse (Maybe RelatedObservations)

Observations related to the problem.

describeProblemObservationsResponse_httpStatus :: Lens' DescribeProblemObservationsResponse Int

The response's http status code.

DescribeObservation

describeObservation_observationId :: Lens' DescribeObservation Text

The ID of the observation.

describeObservationResponse_observation :: Lens' DescribeObservationResponse (Maybe Observation)

Information about the observation.

describeObservationResponse_httpStatus :: Lens' DescribeObservationResponse Int

The response's http status code.

ListLogPatternSets

listLogPatternSets_nextToken :: Lens' ListLogPatternSets (Maybe Text)

The token to request the next page of results.

listLogPatternSets_maxResults :: Lens' ListLogPatternSets (Maybe Natural)

The maximum number of results to return in a single call. To retrieve the remaining results, make another call with the returned NextToken value.

listLogPatternSets_resourceGroupName :: Lens' ListLogPatternSets Text

The name of the resource group.

listLogPatternSetsResponse_resourceGroupName :: Lens' ListLogPatternSetsResponse (Maybe Text)

The name of the resource group.

listLogPatternSetsResponse_nextToken :: Lens' ListLogPatternSetsResponse (Maybe Text)

The token used to retrieve the next page of results. This value is null when there are no more results to return.

listLogPatternSetsResponse_logPatternSets :: Lens' ListLogPatternSetsResponse (Maybe [Text])

The list of log pattern sets.

listLogPatternSetsResponse_httpStatus :: Lens' ListLogPatternSetsResponse Int

The response's http status code.

DescribeComponentConfiguration

describeComponentConfiguration_resourceGroupName :: Lens' DescribeComponentConfiguration Text

The name of the resource group.

describeComponentConfiguration_componentName :: Lens' DescribeComponentConfiguration Text

The name of the component.

describeComponentConfigurationResponse_componentConfiguration :: Lens' DescribeComponentConfigurationResponse (Maybe Text)

The configuration settings of the component. The value is the escaped JSON of the configuration.

describeComponentConfigurationResponse_monitor :: Lens' DescribeComponentConfigurationResponse (Maybe Bool)

Indicates whether the application component is monitored.

describeComponentConfigurationResponse_tier :: Lens' DescribeComponentConfigurationResponse (Maybe Tier)

The tier of the application component. Supported tiers include DOT_NET_CORE, DOT_NET_WORKER, DOT_NET_WEB, SQL_SERVER, and DEFAULT

describeComponentConfigurationResponse_httpStatus :: Lens' DescribeComponentConfigurationResponse Int

The response's http status code.

ListProblems

listProblems_resourceGroupName :: Lens' ListProblems (Maybe Text)

The name of the resource group.

listProblems_startTime :: Lens' ListProblems (Maybe UTCTime)

The time when the problem was detected, in epoch seconds. If you don't specify a time frame for the request, problems within the past seven days are returned.

listProblems_nextToken :: Lens' ListProblems (Maybe Text)

The token to request the next page of results.

listProblems_endTime :: Lens' ListProblems (Maybe UTCTime)

The time when the problem ended, in epoch seconds. If not specified, problems within the past seven days are returned.

listProblems_maxResults :: Lens' ListProblems (Maybe Natural)

The maximum number of results to return in a single call. To retrieve the remaining results, make another call with the returned NextToken value.

listProblemsResponse_nextToken :: Lens' ListProblemsResponse (Maybe Text)

The token used to retrieve the next page of results. This value is null when there are no more results to return.

listProblemsResponse_problemList :: Lens' ListProblemsResponse (Maybe [Problem])

The list of problems.

listProblemsResponse_httpStatus :: Lens' ListProblemsResponse Int

The response's http status code.

ListLogPatterns

listLogPatterns_nextToken :: Lens' ListLogPatterns (Maybe Text)

The token to request the next page of results.

listLogPatterns_patternSetName :: Lens' ListLogPatterns (Maybe Text)

The name of the log pattern set.

listLogPatterns_maxResults :: Lens' ListLogPatterns (Maybe Natural)

The maximum number of results to return in a single call. To retrieve the remaining results, make another call with the returned NextToken value.

listLogPatterns_resourceGroupName :: Lens' ListLogPatterns Text

The name of the resource group.

listLogPatternsResponse_resourceGroupName :: Lens' ListLogPatternsResponse (Maybe Text)

The name of the resource group.

listLogPatternsResponse_nextToken :: Lens' ListLogPatternsResponse (Maybe Text)

The token used to retrieve the next page of results. This value is null when there are no more results to return.

listLogPatternsResponse_logPatterns :: Lens' ListLogPatternsResponse (Maybe [LogPattern])

The list of log patterns.

listLogPatternsResponse_httpStatus :: Lens' ListLogPatternsResponse Int

The response's http status code.

DeleteLogPattern

deleteLogPattern_resourceGroupName :: Lens' DeleteLogPattern Text

The name of the resource group.

deleteLogPattern_patternSetName :: Lens' DeleteLogPattern Text

The name of the log pattern set.

deleteLogPattern_patternName :: Lens' DeleteLogPattern Text

The name of the log pattern.

deleteLogPatternResponse_httpStatus :: Lens' DeleteLogPatternResponse Int

The response's http status code.

UpdateLogPattern

updateLogPattern_pattern :: Lens' UpdateLogPattern (Maybe Text)

The log pattern. The pattern must be DFA compatible. Patterns that utilize forward lookahead or backreference constructions are not supported.

updateLogPattern_rank :: Lens' UpdateLogPattern (Maybe Int)

Rank of the log pattern. Must be a value between 1 and 1,000,000. The patterns are sorted by rank, so we recommend that you set your highest priority patterns with the lowest rank. A pattern of rank 1 will be the first to get matched to a log line. A pattern of rank 1,000,000 will be last to get matched. When you configure custom log patterns from the console, a Low severity pattern translates to a 750,000 rank. A Medium severity pattern translates to a 500,000 rank. And a High severity pattern translates to a 250,000 rank. Rank values less than 1 or greater than 1,000,000 are reserved for AWS-provided patterns.

updateLogPattern_resourceGroupName :: Lens' UpdateLogPattern Text

The name of the resource group.

updateLogPattern_patternSetName :: Lens' UpdateLogPattern Text

The name of the log pattern set.

updateLogPattern_patternName :: Lens' UpdateLogPattern Text

The name of the log pattern.

updateLogPatternResponse_logPattern :: Lens' UpdateLogPatternResponse (Maybe LogPattern)

The successfully created log pattern.

updateLogPatternResponse_resourceGroupName :: Lens' UpdateLogPatternResponse (Maybe Text)

The name of the resource group.

updateLogPatternResponse_httpStatus :: Lens' UpdateLogPatternResponse Int

The response's http status code.

CreateLogPattern

createLogPattern_resourceGroupName :: Lens' CreateLogPattern Text

The name of the resource group.

createLogPattern_patternSetName :: Lens' CreateLogPattern Text

The name of the log pattern set.

createLogPattern_patternName :: Lens' CreateLogPattern Text

The name of the log pattern.

createLogPattern_pattern :: Lens' CreateLogPattern Text

The log pattern. The pattern must be DFA compatible. Patterns that utilize forward lookahead or backreference constructions are not supported.

createLogPattern_rank :: Lens' CreateLogPattern Int

Rank of the log pattern. Must be a value between 1 and 1,000,000. The patterns are sorted by rank, so we recommend that you set your highest priority patterns with the lowest rank. A pattern of rank 1 will be the first to get matched to a log line. A pattern of rank 1,000,000 will be last to get matched. When you configure custom log patterns from the console, a Low severity pattern translates to a 750,000 rank. A Medium severity pattern translates to a 500,000 rank. And a High severity pattern translates to a 250,000 rank. Rank values less than 1 or greater than 1,000,000 are reserved for AWS-provided patterns.

createLogPatternResponse_logPattern :: Lens' CreateLogPatternResponse (Maybe LogPattern)

The successfully created log pattern.

createLogPatternResponse_resourceGroupName :: Lens' CreateLogPatternResponse (Maybe Text)

The name of the resource group.

createLogPatternResponse_httpStatus :: Lens' CreateLogPatternResponse Int

The response's http status code.

ListConfigurationHistory

listConfigurationHistory_resourceGroupName :: Lens' ListConfigurationHistory (Maybe Text)

Resource group to which the application belongs.

listConfigurationHistory_startTime :: Lens' ListConfigurationHistory (Maybe UTCTime)

The start time of the event.

listConfigurationHistory_eventStatus :: Lens' ListConfigurationHistory (Maybe ConfigurationEventStatus)

The status of the configuration update event. Possible values include INFO, WARN, and ERROR.

listConfigurationHistory_nextToken :: Lens' ListConfigurationHistory (Maybe Text)

The NextToken value returned from a previous paginated ListConfigurationHistory request where MaxResults was used and the results exceeded the value of that parameter. Pagination continues from the end of the previous results that returned the NextToken value. This value is null when there are no more results to return.

listConfigurationHistory_endTime :: Lens' ListConfigurationHistory (Maybe UTCTime)

The end time of the event.

listConfigurationHistory_maxResults :: Lens' ListConfigurationHistory (Maybe Natural)

The maximum number of results returned by ListConfigurationHistory in paginated output. When this parameter is used, ListConfigurationHistory returns only MaxResults in a single page along with a NextToken response element. The remaining results of the initial request can be seen by sending another ListConfigurationHistory request with the returned NextToken value. If this parameter is not used, then ListConfigurationHistory returns all results.

listConfigurationHistoryResponse_nextToken :: Lens' ListConfigurationHistoryResponse (Maybe Text)

The NextToken value to include in a future ListConfigurationHistory request. When the results of a ListConfigurationHistory request exceed MaxResults, this value can be used to retrieve the next page of results. This value is null when there are no more results to return.

listConfigurationHistoryResponse_eventList :: Lens' ListConfigurationHistoryResponse (Maybe [ConfigurationEvent])

The list of configuration events and their corresponding details.

listConfigurationHistoryResponse_httpStatus :: Lens' ListConfigurationHistoryResponse Int

The response's http status code.

TagResource

tagResource_resourceARN :: Lens' TagResource Text

The Amazon Resource Name (ARN) of the application that you want to add one or more tags to.

tagResource_tags :: Lens' TagResource [Tag]

A list of tags that to add to the application. A tag consists of a required tag key (Key) and an associated tag value (Value). The maximum length of a tag key is 128 characters. The maximum length of a tag value is 256 characters.

tagResourceResponse_httpStatus :: Lens' TagResourceResponse Int

The response's http status code.

ListApplications

listApplications_nextToken :: Lens' ListApplications (Maybe Text)

The token to request the next page of results.

listApplications_maxResults :: Lens' ListApplications (Maybe Natural)

The maximum number of results to return in a single call. To retrieve the remaining results, make another call with the returned NextToken value.

listApplicationsResponse_nextToken :: Lens' ListApplicationsResponse (Maybe Text)

The token used to retrieve the next page of results. This value is null when there are no more results to return.

listApplicationsResponse_applicationInfoList :: Lens' ListApplicationsResponse (Maybe [ApplicationInfo])

The list of applications.

listApplicationsResponse_httpStatus :: Lens' ListApplicationsResponse Int

The response's http status code.

UntagResource

untagResource_resourceARN :: Lens' UntagResource Text

The Amazon Resource Name (ARN) of the application that you want to remove one or more tags from.

untagResource_tagKeys :: Lens' UntagResource [Text]

The tags (tag keys) that you want to remove from the resource. When you specify a tag key, the action removes both that key and its associated tag value.

To remove more than one tag from the application, append the TagKeys parameter and argument for each additional tag to remove, separated by an ampersand.

untagResourceResponse_httpStatus :: Lens' UntagResourceResponse Int

The response's http status code.

CreateComponent

createComponent_resourceGroupName :: Lens' CreateComponent Text

The name of the resource group.

createComponent_componentName :: Lens' CreateComponent Text

The name of the component.

createComponent_resourceList :: Lens' CreateComponent [Text]

The list of resource ARNs that belong to the component.

createComponentResponse_httpStatus :: Lens' CreateComponentResponse Int

The response's http status code.

ListComponents

listComponents_nextToken :: Lens' ListComponents (Maybe Text)

The token to request the next page of results.

listComponents_maxResults :: Lens' ListComponents (Maybe Natural)

The maximum number of results to return in a single call. To retrieve the remaining results, make another call with the returned NextToken value.

listComponents_resourceGroupName :: Lens' ListComponents Text

The name of the resource group.

listComponentsResponse_applicationComponentList :: Lens' ListComponentsResponse (Maybe [ApplicationComponent])

The list of application components.

listComponentsResponse_nextToken :: Lens' ListComponentsResponse (Maybe Text)

The token to request the next page of results.

listComponentsResponse_httpStatus :: Lens' ListComponentsResponse Int

The response's http status code.

DeleteComponent

deleteComponent_resourceGroupName :: Lens' DeleteComponent Text

The name of the resource group.

deleteComponent_componentName :: Lens' DeleteComponent Text

The name of the component.

deleteComponentResponse_httpStatus :: Lens' DeleteComponentResponse Int

The response's http status code.

UpdateComponent

updateComponent_newComponentName :: Lens' UpdateComponent (Maybe Text)

The new name of the component.

updateComponent_resourceList :: Lens' UpdateComponent (Maybe [Text])

The list of resource ARNs that belong to the component.

updateComponent_resourceGroupName :: Lens' UpdateComponent Text

The name of the resource group.

updateComponent_componentName :: Lens' UpdateComponent Text

The name of the component.

updateComponentResponse_httpStatus :: Lens' UpdateComponentResponse Int

The response's http status code.

DescribeLogPattern

describeLogPattern_resourceGroupName :: Lens' DescribeLogPattern Text

The name of the resource group.

describeLogPattern_patternSetName :: Lens' DescribeLogPattern Text

The name of the log pattern set.

describeLogPattern_patternName :: Lens' DescribeLogPattern Text

The name of the log pattern.

describeLogPatternResponse_logPattern :: Lens' DescribeLogPatternResponse (Maybe LogPattern)

The successfully created log pattern.

describeLogPatternResponse_resourceGroupName :: Lens' DescribeLogPatternResponse (Maybe Text)

The name of the resource group.

describeLogPatternResponse_httpStatus :: Lens' DescribeLogPatternResponse Int

The response's http status code.

Types

ApplicationComponent

applicationComponent_osType :: Lens' ApplicationComponent (Maybe OsType)

The operating system of the component.

applicationComponent_resourceType :: Lens' ApplicationComponent (Maybe Text)

The resource type. Supported resource types include EC2 instances, Auto Scaling group, Classic ELB, Application ELB, and SQS Queue.

applicationComponent_detectedWorkload :: Lens' ApplicationComponent (Maybe (HashMap Tier (HashMap Text Text)))

Workloads detected in the application component.

applicationComponent_monitor :: Lens' ApplicationComponent (Maybe Bool)

Indicates whether the application component is monitored.

applicationComponent_tier :: Lens' ApplicationComponent (Maybe Tier)

The stack tier of the application component.

applicationComponent_componentName :: Lens' ApplicationComponent (Maybe Text)

The name of the component.

applicationComponent_componentRemarks :: Lens' ApplicationComponent (Maybe Text)

If logging is supported for the resource type, indicates whether the component has configured logs to be monitored.

ApplicationInfo

applicationInfo_resourceGroupName :: Lens' ApplicationInfo (Maybe Text)

The name of the resource group used for the application.

applicationInfo_cWEMonitorEnabled :: Lens' ApplicationInfo (Maybe Bool)

Indicates whether Application Insights can listen to CloudWatch events for the application resources, such as instance terminated, failed deployment, and others.

applicationInfo_opsItemSNSTopicArn :: Lens' ApplicationInfo (Maybe Text)

The SNS topic provided to Application Insights that is associated to the created opsItems to receive SNS notifications for opsItem updates.

applicationInfo_lifeCycle :: Lens' ApplicationInfo (Maybe Text)

The lifecycle of the application.

applicationInfo_opsCenterEnabled :: Lens' ApplicationInfo (Maybe Bool)

Indicates whether Application Insights will create opsItems for any problem detected by Application Insights for an application.

applicationInfo_remarks :: Lens' ApplicationInfo (Maybe Text)

The issues on the user side that block Application Insights from successfully monitoring an application. Example remarks include:

“Configuring application, detected 1 Errors, 3 Warnings”

“Configuring application, detected 1 Unconfigured Components”

ConfigurationEvent

configurationEvent_monitoredResourceARN :: Lens' ConfigurationEvent (Maybe Text)

The resource monitored by Application Insights.

configurationEvent_eventStatus :: Lens' ConfigurationEvent (Maybe ConfigurationEventStatus)

The status of the configuration update event. Possible values include INFO, WARN, and ERROR.

configurationEvent_eventResourceName :: Lens' ConfigurationEvent (Maybe Text)

The name of the resource Application Insights attempted to configure.

configurationEvent_eventTime :: Lens' ConfigurationEvent (Maybe UTCTime)

The timestamp of the event.

configurationEvent_eventDetail :: Lens' ConfigurationEvent (Maybe Text)

The details of the event in plain text.

configurationEvent_eventResourceType :: Lens' ConfigurationEvent (Maybe ConfigurationEventResourceType)

The resource type that Application Insights attempted to configure, for example, CLOUDWATCH_ALARM.

LogPattern

logPattern_pattern :: Lens' LogPattern (Maybe Text)

A regular expression that defines the log pattern. A log pattern can contain as many as 50 characters, and it cannot be empty. The pattern must be DFA compatible. Patterns that utilize forward lookahead or backreference constructions are not supported.

logPattern_patternName :: Lens' LogPattern (Maybe Text)

The name of the log pattern. A log pattern name can contain as many as 50 characters, and it cannot be empty. The characters can be Unicode letters, digits, or one of the following symbols: period, dash, underscore.

logPattern_patternSetName :: Lens' LogPattern (Maybe Text)

The name of the log pattern. A log pattern name can contain as many as 30 characters, and it cannot be empty. The characters can be Unicode letters, digits, or one of the following symbols: period, dash, underscore.

logPattern_rank :: Lens' LogPattern (Maybe Int)

Rank of the log pattern. Must be a value between 1 and 1,000,000. The patterns are sorted by rank, so we recommend that you set your highest priority patterns with the lowest rank. A pattern of rank 1 will be the first to get matched to a log line. A pattern of rank 1,000,000 will be last to get matched. When you configure custom log patterns from the console, a Low severity pattern translates to a 750,000 rank. A Medium severity pattern translates to a 500,000 rank. And a High severity pattern translates to a 250,000 rank. Rank values less than 1 or greater than 1,000,000 are reserved for AWS-provided patterns.

Observation

observation_codeDeployApplication :: Lens' Observation (Maybe Text)

The CodeDeploy application to which the deployment belongs.

observation_rdsEventMessage :: Lens' Observation (Maybe Text)

The message of an RDS event.

observation_codeDeployDeploymentId :: Lens' Observation (Maybe Text)

The deployment ID of the CodeDeploy-based observation related to the detected problem.

observation_startTime :: Lens' Observation (Maybe UTCTime)

The time when the observation was first detected, in epoch seconds.

observation_sourceType :: Lens' Observation (Maybe Text)

The source type of the observation.

observation_sourceARN :: Lens' Observation (Maybe Text)

The source resource ARN of the observation.

observation_xRayRequestAverageLatency :: Lens' Observation (Maybe Integer)

The X-Ray node request average latency for this node.

observation_statesStatus :: Lens' Observation (Maybe Text)

The status of the step function-related observation.

observation_codeDeployDeploymentGroup :: Lens' Observation (Maybe Text)

The deployment group to which the CodeDeploy deployment belongs.

observation_healthEventTypeCategory :: Lens' Observation (Maybe Text)

The category of the AWS Health event, such as issue.

observation_xRayRequestCount :: Lens' Observation (Maybe Int)

The X-Ray request count for this node.

observation_s3EventName :: Lens' Observation (Maybe Text)

The name of the S3 CloudWatch Event-based observation.

observation_metricName :: Lens' Observation (Maybe Text)

The name of the observation metric.

observation_ec2State :: Lens' Observation (Maybe Text)

The state of the instance, such as STOPPING or TERMINATING.

observation_logGroup :: Lens' Observation (Maybe Text)

The log group name.

observation_value :: Lens' Observation (Maybe Double)

The value of the source observation metric.

observation_healthEventDescription :: Lens' Observation (Maybe Text)

The description of the AWS Health event provided by the service, such as Amazon EC2.

observation_cloudWatchEventSource :: Lens' Observation (Maybe CloudWatchEventSource)

The source of the CloudWatch Event.

observation_codeDeployState :: Lens' Observation (Maybe Text)

The status of the CodeDeploy deployment, for example SUCCESS or FAILURE.

observation_xRayErrorPercent :: Lens' Observation (Maybe Int)

The X-Ray request error percentage for this node.

observation_statesArn :: Lens' Observation (Maybe Text)

The Amazon Resource Name (ARN) of the step function-based observation.

observation_cloudWatchEventId :: Lens' Observation (Maybe Text)

The ID of the CloudWatch Event-based observation related to the detected problem.

observation_logText :: Lens' Observation (Maybe Text)

The log text of the observation.

observation_logFilter :: Lens' Observation (Maybe LogFilter)

The log filter of the observation.

observation_metricNamespace :: Lens' Observation (Maybe Text)

The namespace of the observation metric.

observation_rdsEventCategories :: Lens' Observation (Maybe Text)

The category of an RDS event.

observation_xRayNodeType :: Lens' Observation (Maybe Text)

The type of the X-Ray node.

observation_endTime :: Lens' Observation (Maybe UTCTime)

The time when the observation ended, in epoch seconds.

observation_statesInput :: Lens' Observation (Maybe Text)

The input to the step function-based observation.

observation_xRayNodeName :: Lens' Observation (Maybe Text)

The name of the X-Ray node.

observation_id :: Lens' Observation (Maybe Text)

The ID of the observation type.

observation_healthEventArn :: Lens' Observation (Maybe Text)

The Amazon Resource Name (ARN) of the AWS Health Event-based observation.

observation_healthEventTypeCode :: Lens' Observation (Maybe Text)

The type of the AWS Health event, for example, AWS_EC2_POWER_CONNECTIVITY_ISSUE.

observation_ebsResult :: Lens' Observation (Maybe Text)

The result of an EBS CloudWatch event, such as failed or succeeded.

observation_cloudWatchEventDetailType :: Lens' Observation (Maybe Text)

The detail type of the CloudWatch Event-based observation, for example, EC2 Instance State-change Notification.

observation_codeDeployInstanceGroupId :: Lens' Observation (Maybe Text)

The instance group to which the CodeDeploy instance belongs.

observation_ebsCause :: Lens' Observation (Maybe Text)

The cause of an EBS CloudWatch event.

observation_ebsEvent :: Lens' Observation (Maybe Text)

The type of EBS CloudWatch event, such as createVolume, deleteVolume or attachVolume.

observation_ebsRequestId :: Lens' Observation (Maybe Text)

The request ID of an EBS CloudWatch event.

observation_xRayFaultPercent :: Lens' Observation (Maybe Int)

The X-Ray request fault percentage for this node.

observation_statesExecutionArn :: Lens' Observation (Maybe Text)

The Amazon Resource Name (ARN) of the step function execution-based observation.

observation_lineTime :: Lens' Observation (Maybe UTCTime)

The timestamp in the CloudWatch Logs that specifies when the matched line occurred.

observation_unit :: Lens' Observation (Maybe Text)

The unit of the source observation metric.

observation_xRayThrottlePercent :: Lens' Observation (Maybe Int)

The X-Ray request throttle percentage for this node.

observation_healthService :: Lens' Observation (Maybe Text)

The service to which the AWS Health Event belongs, such as EC2.

Problem

problem_status :: Lens' Problem (Maybe Status)

The status of the problem.

problem_resourceGroupName :: Lens' Problem (Maybe Text)

The name of the resource group affected by the problem.

problem_startTime :: Lens' Problem (Maybe UTCTime)

The time when the problem started, in epoch seconds.

problem_insights :: Lens' Problem (Maybe Text)

A detailed analysis of the problem using machine learning.

problem_endTime :: Lens' Problem (Maybe UTCTime)

The time when the problem ended, in epoch seconds.

problem_id :: Lens' Problem (Maybe Text)

The ID of the problem.

problem_severityLevel :: Lens' Problem (Maybe SeverityLevel)

A measure of the level of impact of the problem.

problem_title :: Lens' Problem (Maybe Text)

The name of the problem.

problem_affectedResource :: Lens' Problem (Maybe Text)

The resource affected by the problem.

problem_feedback :: Lens' Problem (Maybe (HashMap FeedbackKey FeedbackValue))

Feedback provided by the user about the problem.

RelatedObservations

relatedObservations_observationList :: Lens' RelatedObservations (Maybe [Observation])

The list of observations related to the problem.

Tag

tag_key :: Lens' Tag Text

One part of a key-value pair that defines a tag. The maximum length of a tag key is 128 characters. The minimum length is 1 character.

tag_value :: Lens' Tag Text

The optional part of a key-value pair that defines a tag. The maximum length of a tag value is 256 characters. The minimum length is 0 characters. If you don't want an application to have a specific tag value, don't specify a value for this parameter.