Amazonka.ApplicationInsights.Types.Problem

Description

Documentation

data Problem

Describes a problem that is detected by correlating observations.

See: newProblem smart constructor.

Constructors

Problem'

Fields

status :: Maybe Status

The status of the problem.

resourceGroupName :: Maybe Text

The name of the resource group affected by the problem.

startTime :: Maybe POSIX

The time when the problem started, in epoch seconds.

insights :: Maybe Text

A detailed analysis of the problem using machine learning.

endTime :: Maybe POSIX

The time when the problem ended, in epoch seconds.

id :: Maybe Text

The ID of the problem.

severityLevel :: Maybe SeverityLevel

A measure of the level of impact of the problem.

title :: Maybe Text

The name of the problem.

affectedResource :: Maybe Text

The resource affected by the problem.

feedback :: Maybe (HashMap FeedbackKey FeedbackValue)

Feedback provided by the user about the problem.

newProblem :: Problem

Create a value of Problem with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:status:Problem', problem_status - The status of the problem.

$sel:resourceGroupName:Problem', problem_resourceGroupName - The name of the resource group affected by the problem.

$sel:startTime:Problem', problem_startTime - The time when the problem started, in epoch seconds.

$sel:insights:Problem', problem_insights - A detailed analysis of the problem using machine learning.

$sel:endTime:Problem', problem_endTime - The time when the problem ended, in epoch seconds.

$sel:id:Problem', problem_id - The ID of the problem.

$sel:severityLevel:Problem', problem_severityLevel - A measure of the level of impact of the problem.

$sel:title:Problem', problem_title - The name of the problem.

$sel:affectedResource:Problem', problem_affectedResource - The resource affected by the problem.

$sel:feedback:Problem', problem_feedback - Feedback provided by the user about the problem.

problem_status :: Lens' Problem (Maybe Status)

The status of the problem.

problem_resourceGroupName :: Lens' Problem (Maybe Text)

The name of the resource group affected by the problem.

problem_startTime :: Lens' Problem (Maybe UTCTime)

The time when the problem started, in epoch seconds.

problem_insights :: Lens' Problem (Maybe Text)

A detailed analysis of the problem using machine learning.

problem_endTime :: Lens' Problem (Maybe UTCTime)

The time when the problem ended, in epoch seconds.

problem_id :: Lens' Problem (Maybe Text)

The ID of the problem.

problem_severityLevel :: Lens' Problem (Maybe SeverityLevel)

A measure of the level of impact of the problem.

problem_title :: Lens' Problem (Maybe Text)

The name of the problem.

problem_affectedResource :: Lens' Problem (Maybe Text)

The resource affected by the problem.

problem_feedback :: Lens' Problem (Maybe (HashMap FeedbackKey FeedbackValue))

Feedback provided by the user about the problem.