Amazonka.ApplicationInsights.UpdateComponentConfiguration

Description

Updates the monitoring configurations for the component. The configuration input parameter is an escaped JSON of the configuration and should match the schema of what is returned by DescribeComponentConfigurationRecommendation.

Creating a Request

data UpdateComponentConfiguration

See: newUpdateComponentConfiguration smart constructor.

Constructors

UpdateComponentConfiguration'

Fields

componentConfiguration :: Maybe Text

The configuration settings of the component. The value is the escaped JSON of the configuration. For more information about the JSON format, see Working with JSON. You can send a request to DescribeComponentConfigurationRecommendation to see the recommended configuration for a component. For the complete format of the component configuration file, see Component Configuration.

monitor :: Maybe Bool

Indicates whether the application component is monitored.

tier :: Maybe Tier

The tier of the application component. Supported tiers include DOT_NET_WORKER, DOT_NET_WEB, DOT_NET_CORE, SQL_SERVER, and DEFAULT.

resourceGroupName :: Text

The name of the resource group.

componentName :: Text

The name of the component.

newUpdateComponentConfiguration

Arguments

:: Text
$sel:resourceGroupName:UpdateComponentConfiguration'

-> Text
$sel:componentName:UpdateComponentConfiguration'

-> UpdateComponentConfiguration

Create a value of UpdateComponentConfiguration with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:componentConfiguration:UpdateComponentConfiguration', updateComponentConfiguration_componentConfiguration - The configuration settings of the component. The value is the escaped JSON of the configuration. For more information about the JSON format, see Working with JSON. You can send a request to DescribeComponentConfigurationRecommendation to see the recommended configuration for a component. For the complete format of the component configuration file, see Component Configuration.

$sel:monitor:UpdateComponentConfiguration', updateComponentConfiguration_monitor - Indicates whether the application component is monitored.

$sel:tier:UpdateComponentConfiguration', updateComponentConfiguration_tier - The tier of the application component. Supported tiers include DOT_NET_WORKER, DOT_NET_WEB, DOT_NET_CORE, SQL_SERVER, and DEFAULT.

$sel:resourceGroupName:UpdateComponentConfiguration', updateComponentConfiguration_resourceGroupName - The name of the resource group.

$sel:componentName:UpdateComponentConfiguration', updateComponentConfiguration_componentName - The name of the component.

Request Lenses

updateComponentConfiguration_componentConfiguration :: Lens' UpdateComponentConfiguration (Maybe Text)

The configuration settings of the component. The value is the escaped JSON of the configuration. For more information about the JSON format, see Working with JSON. You can send a request to DescribeComponentConfigurationRecommendation to see the recommended configuration for a component. For the complete format of the component configuration file, see Component Configuration.

updateComponentConfiguration_monitor :: Lens' UpdateComponentConfiguration (Maybe Bool)

Indicates whether the application component is monitored.

updateComponentConfiguration_tier :: Lens' UpdateComponentConfiguration (Maybe Tier)

The tier of the application component. Supported tiers include DOT_NET_WORKER, DOT_NET_WEB, DOT_NET_CORE, SQL_SERVER, and DEFAULT.

updateComponentConfiguration_resourceGroupName :: Lens' UpdateComponentConfiguration Text

The name of the resource group.

updateComponentConfiguration_componentName :: Lens' UpdateComponentConfiguration Text

The name of the component.

Destructuring the Response

data UpdateComponentConfigurationResponse

See: newUpdateComponentConfigurationResponse smart constructor.

Constructors

UpdateComponentConfigurationResponse'

Fields

httpStatus :: Int

The response's http status code.

newUpdateComponentConfigurationResponse

Arguments

:: Int
$sel:httpStatus:UpdateComponentConfigurationResponse'

-> UpdateComponentConfigurationResponse

Create a value of UpdateComponentConfigurationResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:httpStatus:UpdateComponentConfigurationResponse', updateComponentConfigurationResponse_httpStatus - The response's http status code.

Response Lenses

updateComponentConfigurationResponse_httpStatus :: Lens' UpdateComponentConfigurationResponse Int

The response's http status code.