Amazonka.Athena.BatchGetNamedQuery

Description

Returns the details of a single named query or a list of up to 50 queries, which you provide as an array of query ID strings. Requires you to have access to the workgroup in which the queries were saved. Use ListNamedQueriesInput to get the list of named query IDs in the specified workgroup. If information could not be retrieved for a submitted query ID, information about the query ID submitted is listed under UnprocessedNamedQueryId. Named queries differ from executed queries. Use BatchGetQueryExecutionInput to get details about each unique query execution, and ListQueryExecutionsInput to get a list of query execution IDs.

Creating a Request

data BatchGetNamedQuery

See: newBatchGetNamedQuery smart constructor.

Constructors

BatchGetNamedQuery'

Fields

namedQueryIds :: NonEmpty Text

An array of query IDs.

newBatchGetNamedQuery

Arguments

:: NonEmpty Text
$sel:namedQueryIds:BatchGetNamedQuery'

-> BatchGetNamedQuery

Create a value of BatchGetNamedQuery with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:namedQueryIds:BatchGetNamedQuery', batchGetNamedQuery_namedQueryIds - An array of query IDs.

Request Lenses

batchGetNamedQuery_namedQueryIds :: Lens' BatchGetNamedQuery (NonEmpty Text)

An array of query IDs.

Destructuring the Response

data BatchGetNamedQueryResponse

See: newBatchGetNamedQueryResponse smart constructor.

Constructors

BatchGetNamedQueryResponse'

Fields

namedQueries :: Maybe [NamedQuery]

Information about the named query IDs submitted.

unprocessedNamedQueryIds :: Maybe [UnprocessedNamedQueryId]

Information about provided query IDs.

httpStatus :: Int

The response's http status code.

newBatchGetNamedQueryResponse

Arguments

:: Int
$sel:httpStatus:BatchGetNamedQueryResponse'

-> BatchGetNamedQueryResponse

Create a value of BatchGetNamedQueryResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:namedQueries:BatchGetNamedQueryResponse', batchGetNamedQueryResponse_namedQueries - Information about the named query IDs submitted.

$sel:unprocessedNamedQueryIds:BatchGetNamedQueryResponse', batchGetNamedQueryResponse_unprocessedNamedQueryIds - Information about provided query IDs.

$sel:httpStatus:BatchGetNamedQueryResponse', batchGetNamedQueryResponse_httpStatus - The response's http status code.

Response Lenses

batchGetNamedQueryResponse_namedQueries :: Lens' BatchGetNamedQueryResponse (Maybe [NamedQuery])

Information about the named query IDs submitted.

batchGetNamedQueryResponse_unprocessedNamedQueryIds :: Lens' BatchGetNamedQueryResponse (Maybe [UnprocessedNamedQueryId])

Information about provided query IDs.

batchGetNamedQueryResponse_httpStatus :: Lens' BatchGetNamedQueryResponse Int

The response's http status code.