Amazonka.Athena.GetNamedQuery

Description

Returns information about a single query. Requires that you have access to the workgroup in which the query was saved.

Creating a Request

data GetNamedQuery

See: newGetNamedQuery smart constructor.

Constructors

GetNamedQuery'

Fields

namedQueryId :: Text

The unique ID of the query. Use ListNamedQueries to get query IDs.

newGetNamedQuery

Arguments

:: Text
$sel:namedQueryId:GetNamedQuery'

-> GetNamedQuery

Create a value of GetNamedQuery with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:namedQueryId:GetNamedQuery', getNamedQuery_namedQueryId - The unique ID of the query. Use ListNamedQueries to get query IDs.

Request Lenses

getNamedQuery_namedQueryId :: Lens' GetNamedQuery Text

The unique ID of the query. Use ListNamedQueries to get query IDs.

Destructuring the Response

data GetNamedQueryResponse

See: newGetNamedQueryResponse smart constructor.

Constructors

GetNamedQueryResponse'

Fields

namedQuery :: Maybe NamedQuery

Information about the query.

httpStatus :: Int

The response's http status code.

newGetNamedQueryResponse

Arguments

:: Int
$sel:httpStatus:GetNamedQueryResponse'

-> GetNamedQueryResponse

Create a value of GetNamedQueryResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:namedQuery:GetNamedQueryResponse', getNamedQueryResponse_namedQuery - Information about the query.

$sel:httpStatus:GetNamedQueryResponse', getNamedQueryResponse_httpStatus - The response's http status code.

Response Lenses

getNamedQueryResponse_namedQuery :: Lens' GetNamedQueryResponse (Maybe NamedQuery)

Information about the query.

getNamedQueryResponse_httpStatus :: Lens' GetNamedQueryResponse Int

The response's http status code.