Amazonka.Athena.ListDataCatalogs

Description

Lists the data catalogs in the current Amazon Web Services account.

This operation returns paginated results.

Creating a Request

data ListDataCatalogs

See: newListDataCatalogs smart constructor.

Constructors

ListDataCatalogs'

Fields

nextToken :: Maybe Text

A token generated by the Athena service that specifies where to continue pagination if a previous request was truncated. To obtain the next set of pages, pass in the NextToken from the response object of the previous page call.

maxResults :: Maybe Natural

Specifies the maximum number of data catalogs to return.

newListDataCatalogs :: ListDataCatalogs

Create a value of ListDataCatalogs with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:nextToken:ListDataCatalogs', listDataCatalogs_nextToken - A token generated by the Athena service that specifies where to continue pagination if a previous request was truncated. To obtain the next set of pages, pass in the NextToken from the response object of the previous page call.

$sel:maxResults:ListDataCatalogs', listDataCatalogs_maxResults - Specifies the maximum number of data catalogs to return.

Request Lenses

listDataCatalogs_nextToken :: Lens' ListDataCatalogs (Maybe Text)

A token generated by the Athena service that specifies where to continue pagination if a previous request was truncated. To obtain the next set of pages, pass in the NextToken from the response object of the previous page call.

listDataCatalogs_maxResults :: Lens' ListDataCatalogs (Maybe Natural)

Specifies the maximum number of data catalogs to return.

Destructuring the Response

data ListDataCatalogsResponse

See: newListDataCatalogsResponse smart constructor.

Constructors

ListDataCatalogsResponse'

Fields

dataCatalogsSummary :: Maybe [DataCatalogSummary]

A summary list of data catalogs.

nextToken :: Maybe Text

A token generated by the Athena service that specifies where to continue pagination if a previous request was truncated. To obtain the next set of pages, pass in the NextToken from the response object of the previous page call.

httpStatus :: Int

The response's http status code.

newListDataCatalogsResponse

Arguments

:: Int
$sel:httpStatus:ListDataCatalogsResponse'

-> ListDataCatalogsResponse

Create a value of ListDataCatalogsResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:dataCatalogsSummary:ListDataCatalogsResponse', listDataCatalogsResponse_dataCatalogsSummary - A summary list of data catalogs.

$sel:nextToken:ListDataCatalogs', listDataCatalogsResponse_nextToken - A token generated by the Athena service that specifies where to continue pagination if a previous request was truncated. To obtain the next set of pages, pass in the NextToken from the response object of the previous page call.

$sel:httpStatus:ListDataCatalogsResponse', listDataCatalogsResponse_httpStatus - The response's http status code.

Response Lenses

listDataCatalogsResponse_dataCatalogsSummary :: Lens' ListDataCatalogsResponse (Maybe [DataCatalogSummary])

A summary list of data catalogs.

listDataCatalogsResponse_nextToken :: Lens' ListDataCatalogsResponse (Maybe Text)

A token generated by the Athena service that specifies where to continue pagination if a previous request was truncated. To obtain the next set of pages, pass in the NextToken from the response object of the previous page call.

listDataCatalogsResponse_httpStatus :: Lens' ListDataCatalogsResponse Int

The response's http status code.