Amazonka.Athena.Types.ColumnInfo

Description

Documentation

data ColumnInfo

Information about the columns in a query execution result.

See: newColumnInfo smart constructor.

Constructors

ColumnInfo'

Fields

scale :: Maybe Int

For DECIMAL data types, specifies the total number of digits in the fractional part of the value. Defaults to 0.

precision :: Maybe Int

For DECIMAL data types, specifies the total number of digits, up to 38. For performance reasons, we recommend up to 18 digits.

schemaName :: Maybe Text

The schema name (database name) to which the query results belong.

catalogName :: Maybe Text

The catalog to which the query results belong.

caseSensitive :: Maybe Bool

Indicates whether values in the column are case-sensitive.

label :: Maybe Text

A column label.

tableName :: Maybe Text

The table name for the query results.

nullable :: Maybe ColumnNullable

Indicates the column's nullable status.

name :: Text

The name of the column.

type' :: Text

The data type of the column.

newColumnInfo

Arguments

:: Text
$sel:name:ColumnInfo'

-> Text
$sel:type':ColumnInfo'

-> ColumnInfo

Create a value of ColumnInfo with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:scale:ColumnInfo', columnInfo_scale - For DECIMAL data types, specifies the total number of digits in the fractional part of the value. Defaults to 0.

$sel:precision:ColumnInfo', columnInfo_precision - For DECIMAL data types, specifies the total number of digits, up to 38. For performance reasons, we recommend up to 18 digits.

$sel:schemaName:ColumnInfo', columnInfo_schemaName - The schema name (database name) to which the query results belong.

$sel:catalogName:ColumnInfo', columnInfo_catalogName - The catalog to which the query results belong.

$sel:caseSensitive:ColumnInfo', columnInfo_caseSensitive - Indicates whether values in the column are case-sensitive.

$sel:label:ColumnInfo', columnInfo_label - A column label.

$sel:tableName:ColumnInfo', columnInfo_tableName - The table name for the query results.

$sel:nullable:ColumnInfo', columnInfo_nullable - Indicates the column's nullable status.

$sel:name:ColumnInfo', columnInfo_name - The name of the column.

$sel:type':ColumnInfo', columnInfo_type - The data type of the column.

columnInfo_scale :: Lens' ColumnInfo (Maybe Int)

For DECIMAL data types, specifies the total number of digits in the fractional part of the value. Defaults to 0.

columnInfo_precision :: Lens' ColumnInfo (Maybe Int)

For DECIMAL data types, specifies the total number of digits, up to 38. For performance reasons, we recommend up to 18 digits.

columnInfo_schemaName :: Lens' ColumnInfo (Maybe Text)

The schema name (database name) to which the query results belong.

columnInfo_catalogName :: Lens' ColumnInfo (Maybe Text)

The catalog to which the query results belong.

columnInfo_caseSensitive :: Lens' ColumnInfo (Maybe Bool)

Indicates whether values in the column are case-sensitive.

columnInfo_label :: Lens' ColumnInfo (Maybe Text)

A column label.

columnInfo_tableName :: Lens' ColumnInfo (Maybe Text)

The table name for the query results.

columnInfo_nullable :: Lens' ColumnInfo (Maybe ColumnNullable)

Indicates the column's nullable status.

columnInfo_name :: Lens' ColumnInfo Text

The name of the column.

columnInfo_type :: Lens' ColumnInfo Text

The data type of the column.