Amazonka.Athena.Types.NamedQuery

Description

Documentation

data NamedQuery

A query, where QueryString is the list of SQL query statements that comprise the query.

See: newNamedQuery smart constructor.

Constructors

NamedQuery'

Fields

namedQueryId :: Maybe Text

The unique identifier of the query.

description :: Maybe Text

The query description.

workGroup :: Maybe Text

The name of the workgroup that contains the named query.

name :: Text

The query name.

database :: Text

The database to which the query belongs.

queryString :: Text

The SQL query statements that comprise the query.

newNamedQuery

Arguments

:: Text
$sel:name:NamedQuery'

-> Text
$sel:database:NamedQuery'

-> Text
$sel:queryString:NamedQuery'

-> NamedQuery

Create a value of NamedQuery with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:namedQueryId:NamedQuery', namedQuery_namedQueryId - The unique identifier of the query.

$sel:description:NamedQuery', namedQuery_description - The query description.

$sel:workGroup:NamedQuery', namedQuery_workGroup - The name of the workgroup that contains the named query.

$sel:name:NamedQuery', namedQuery_name - The query name.

$sel:database:NamedQuery', namedQuery_database - The database to which the query belongs.

$sel:queryString:NamedQuery', namedQuery_queryString - The SQL query statements that comprise the query.

namedQuery_namedQueryId :: Lens' NamedQuery (Maybe Text)

The unique identifier of the query.

namedQuery_description :: Lens' NamedQuery (Maybe Text)

The query description.

namedQuery_workGroup :: Lens' NamedQuery (Maybe Text)

The name of the workgroup that contains the named query.

namedQuery_name :: Lens' NamedQuery Text

The query name.

namedQuery_database :: Lens' NamedQuery Text

The database to which the query belongs.

namedQuery_queryString :: Lens' NamedQuery Text

The SQL query statements that comprise the query.