Amazonka.Athena.Types.TableMetadata

Description

Documentation

data TableMetadata

Contains metadata for a table.

See: newTableMetadata smart constructor.

Constructors

TableMetadata'

Fields

tableType :: Maybe Text

The type of table. In Athena, only EXTERNAL_TABLE is supported.

parameters :: Maybe (HashMap Text Text)

A set of custom key/value pairs for table properties.

columns :: Maybe [Column]

A list of the columns in the table.

lastAccessTime :: Maybe POSIX

The last time the table was accessed.

partitionKeys :: Maybe [Column]

A list of the partition keys in the table.

createTime :: Maybe POSIX

The time that the table was created.

name :: Text

The name of the table.

newTableMetadata

Arguments

:: Text
$sel:name:TableMetadata'

-> TableMetadata

Create a value of TableMetadata with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:tableType:TableMetadata', tableMetadata_tableType - The type of table. In Athena, only EXTERNAL_TABLE is supported.

$sel:parameters:TableMetadata', tableMetadata_parameters - A set of custom key/value pairs for table properties.

$sel:columns:TableMetadata', tableMetadata_columns - A list of the columns in the table.

$sel:lastAccessTime:TableMetadata', tableMetadata_lastAccessTime - The last time the table was accessed.

$sel:partitionKeys:TableMetadata', tableMetadata_partitionKeys - A list of the partition keys in the table.

$sel:createTime:TableMetadata', tableMetadata_createTime - The time that the table was created.

$sel:name:TableMetadata', tableMetadata_name - The name of the table.

tableMetadata_tableType :: Lens' TableMetadata (Maybe Text)

The type of table. In Athena, only EXTERNAL_TABLE is supported.

tableMetadata_parameters :: Lens' TableMetadata (Maybe (HashMap Text Text))

A set of custom key/value pairs for table properties.

tableMetadata_columns :: Lens' TableMetadata (Maybe [Column])

A list of the columns in the table.

tableMetadata_lastAccessTime :: Lens' TableMetadata (Maybe UTCTime)

The last time the table was accessed.

tableMetadata_partitionKeys :: Lens' TableMetadata (Maybe [Column])

A list of the partition keys in the table.

tableMetadata_createTime :: Lens' TableMetadata (Maybe UTCTime)

The time that the table was created.

tableMetadata_name :: Lens' TableMetadata Text

The name of the table.