Amazonka.Athena.Types.UnprocessedNamedQueryId

Description

Documentation

data UnprocessedNamedQueryId

Information about a named query ID that could not be processed.

See: newUnprocessedNamedQueryId smart constructor.

Constructors

UnprocessedNamedQueryId'

Fields

namedQueryId :: Maybe Text

The unique identifier of the named query.

errorCode :: Maybe Text

The error code returned when the processing request for the named query failed, if applicable.

errorMessage :: Maybe Text

The error message returned when the processing request for the named query failed, if applicable.

newUnprocessedNamedQueryId :: UnprocessedNamedQueryId

Create a value of UnprocessedNamedQueryId with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:namedQueryId:UnprocessedNamedQueryId', unprocessedNamedQueryId_namedQueryId - The unique identifier of the named query.

$sel:errorCode:UnprocessedNamedQueryId', unprocessedNamedQueryId_errorCode - The error code returned when the processing request for the named query failed, if applicable.

$sel:errorMessage:UnprocessedNamedQueryId', unprocessedNamedQueryId_errorMessage - The error message returned when the processing request for the named query failed, if applicable.

unprocessedNamedQueryId_namedQueryId :: Lens' UnprocessedNamedQueryId (Maybe Text)

The unique identifier of the named query.

unprocessedNamedQueryId_errorCode :: Lens' UnprocessedNamedQueryId (Maybe Text)

The error code returned when the processing request for the named query failed, if applicable.

unprocessedNamedQueryId_errorMessage :: Lens' UnprocessedNamedQueryId (Maybe Text)

The error message returned when the processing request for the named query failed, if applicable.