Amazonka.Athena.UpdateDataCatalog

Description

Updates the data catalog that has the specified name.

Creating a Request

data UpdateDataCatalog

See: newUpdateDataCatalog smart constructor.

Constructors

UpdateDataCatalog'

Fields

parameters :: Maybe (HashMap Text Text)

Specifies the Lambda function or functions to use for updating the data catalog. This is a mapping whose values depend on the catalog type.

For the HIVE data catalog type, use the following syntax. The metadata-function parameter is required. The sdk-version parameter is optional and defaults to the currently supported version.

metadata-function=lambda_arn, sdk-version=version_number

For the LAMBDA data catalog type, use one of the following sets of required parameters, but not both.

If you have one Lambda function that processes metadata and another for reading the actual data, use the following syntax. Both parameters are required.

metadata-function=lambda_arn, record-function=lambda_arn

If you have a composite Lambda function that processes both metadata and data, use the following syntax to specify your Lambda function.

function=lambda_arn

description :: Maybe Text

New or modified text that describes the data catalog.

name :: Text

The name of the data catalog to update. The catalog name must be unique for the Amazon Web Services account and can use a maximum of 128 alphanumeric, underscore, at sign, or hyphen characters.

type' :: DataCatalogType

Specifies the type of data catalog to update. Specify LAMBDA for a federated catalog, HIVE for an external hive metastore, or GLUE for an Glue Data Catalog.

newUpdateDataCatalog

Arguments

:: Text
$sel:name:UpdateDataCatalog'

-> DataCatalogType
$sel:type':UpdateDataCatalog'

-> UpdateDataCatalog

Create a value of UpdateDataCatalog with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:parameters:UpdateDataCatalog', updateDataCatalog_parameters - Specifies the Lambda function or functions to use for updating the data catalog. This is a mapping whose values depend on the catalog type.

For the HIVE data catalog type, use the following syntax. The metadata-function parameter is required. The sdk-version parameter is optional and defaults to the currently supported version.

metadata-function=lambda_arn, sdk-version=version_number

For the LAMBDA data catalog type, use one of the following sets of required parameters, but not both.

If you have one Lambda function that processes metadata and another for reading the actual data, use the following syntax. Both parameters are required.

metadata-function=lambda_arn, record-function=lambda_arn

If you have a composite Lambda function that processes both metadata and data, use the following syntax to specify your Lambda function.

function=lambda_arn

$sel:description:UpdateDataCatalog', updateDataCatalog_description - New or modified text that describes the data catalog.

$sel:name:UpdateDataCatalog', updateDataCatalog_name - The name of the data catalog to update. The catalog name must be unique for the Amazon Web Services account and can use a maximum of 128 alphanumeric, underscore, at sign, or hyphen characters.

$sel:type':UpdateDataCatalog', updateDataCatalog_type - Specifies the type of data catalog to update. Specify LAMBDA for a federated catalog, HIVE for an external hive metastore, or GLUE for an Glue Data Catalog.

Request Lenses

updateDataCatalog_parameters :: Lens' UpdateDataCatalog (Maybe (HashMap Text Text))

Specifies the Lambda function or functions to use for updating the data catalog. This is a mapping whose values depend on the catalog type.

For the HIVE data catalog type, use the following syntax. The metadata-function parameter is required. The sdk-version parameter is optional and defaults to the currently supported version.

metadata-function=lambda_arn, sdk-version=version_number

For the LAMBDA data catalog type, use one of the following sets of required parameters, but not both.

If you have one Lambda function that processes metadata and another for reading the actual data, use the following syntax. Both parameters are required.

metadata-function=lambda_arn, record-function=lambda_arn

If you have a composite Lambda function that processes both metadata and data, use the following syntax to specify your Lambda function.

function=lambda_arn

updateDataCatalog_description :: Lens' UpdateDataCatalog (Maybe Text)

New or modified text that describes the data catalog.

updateDataCatalog_name :: Lens' UpdateDataCatalog Text

The name of the data catalog to update. The catalog name must be unique for the Amazon Web Services account and can use a maximum of 128 alphanumeric, underscore, at sign, or hyphen characters.

updateDataCatalog_type :: Lens' UpdateDataCatalog DataCatalogType

Specifies the type of data catalog to update. Specify LAMBDA for a federated catalog, HIVE for an external hive metastore, or GLUE for an Glue Data Catalog.

Destructuring the Response

data UpdateDataCatalogResponse

See: newUpdateDataCatalogResponse smart constructor.

Constructors

UpdateDataCatalogResponse'

Fields

httpStatus :: Int

The response's http status code.

newUpdateDataCatalogResponse

Arguments

:: Int
$sel:httpStatus:UpdateDataCatalogResponse'

-> UpdateDataCatalogResponse

Create a value of UpdateDataCatalogResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:httpStatus:UpdateDataCatalogResponse', updateDataCatalogResponse_httpStatus - The response's http status code.

Response Lenses

updateDataCatalogResponse_httpStatus :: Lens' UpdateDataCatalogResponse Int

The response's http status code.