Amazonka.AuditManager.BatchCreateDelegationByAssessment

Description

Create a batch of delegations for a specified assessment in Audit Manager.

Creating a Request

data BatchCreateDelegationByAssessment

See: newBatchCreateDelegationByAssessment smart constructor.

Constructors

BatchCreateDelegationByAssessment'

Fields

createDelegationRequests :: NonEmpty CreateDelegationRequest

The API request to batch create delegations in Audit Manager.

assessmentId :: Text

The identifier for the specified assessment.

newBatchCreateDelegationByAssessment

Arguments

:: NonEmpty CreateDelegationRequest
$sel:createDelegationRequests:BatchCreateDelegationByAssessment'

-> Text
$sel:assessmentId:BatchCreateDelegationByAssessment'

-> BatchCreateDelegationByAssessment

Create a value of BatchCreateDelegationByAssessment with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:createDelegationRequests:BatchCreateDelegationByAssessment', batchCreateDelegationByAssessment_createDelegationRequests - The API request to batch create delegations in Audit Manager.

$sel:assessmentId:BatchCreateDelegationByAssessment', batchCreateDelegationByAssessment_assessmentId - The identifier for the specified assessment.

Request Lenses

batchCreateDelegationByAssessment_createDelegationRequests :: Lens' BatchCreateDelegationByAssessment (NonEmpty CreateDelegationRequest)

The API request to batch create delegations in Audit Manager.

batchCreateDelegationByAssessment_assessmentId :: Lens' BatchCreateDelegationByAssessment Text

The identifier for the specified assessment.

Destructuring the Response

data BatchCreateDelegationByAssessmentResponse

See: newBatchCreateDelegationByAssessmentResponse smart constructor.

Constructors

BatchCreateDelegationByAssessmentResponse'

Fields

delegations :: Maybe [Delegation]

The delegations associated with the assessment.

errors :: Maybe [BatchCreateDelegationByAssessmentError]

A list of errors returned by the BatchCreateDelegationByAssessment API.

httpStatus :: Int

The response's http status code.

newBatchCreateDelegationByAssessmentResponse

Arguments

:: Int
$sel:httpStatus:BatchCreateDelegationByAssessmentResponse'

-> BatchCreateDelegationByAssessmentResponse

Create a value of BatchCreateDelegationByAssessmentResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:delegations:BatchCreateDelegationByAssessmentResponse', batchCreateDelegationByAssessmentResponse_delegations - The delegations associated with the assessment.

$sel:errors:BatchCreateDelegationByAssessmentResponse', batchCreateDelegationByAssessmentResponse_errors - A list of errors returned by the BatchCreateDelegationByAssessment API.

$sel:httpStatus:BatchCreateDelegationByAssessmentResponse', batchCreateDelegationByAssessmentResponse_httpStatus - The response's http status code.

Response Lenses

batchCreateDelegationByAssessmentResponse_delegations :: Lens' BatchCreateDelegationByAssessmentResponse (Maybe [Delegation])

The delegations associated with the assessment.

batchCreateDelegationByAssessmentResponse_errors :: Lens' BatchCreateDelegationByAssessmentResponse (Maybe [BatchCreateDelegationByAssessmentError])

A list of errors returned by the BatchCreateDelegationByAssessment API.

batchCreateDelegationByAssessmentResponse_httpStatus :: Lens' BatchCreateDelegationByAssessmentResponse Int

The response's http status code.