Amazonka.AuditManager.CreateAssessment

Description

Creates an assessment in Audit Manager.

Creating a Request

data CreateAssessment

See: newCreateAssessment smart constructor.

Constructors

CreateAssessment'

Fields

description :: Maybe Text

The optional description of the assessment to be created.

tags :: Maybe (HashMap Text Text)

The tags associated with the assessment.

name :: Text

The name of the assessment to be created.

assessmentReportsDestination :: AssessmentReportsDestination

The assessment report storage destination for the specified assessment that is being created.

scope :: Scope

roles :: [Role]

The list of roles for the specified assessment.

frameworkId :: Text

The identifier for the specified framework.

newCreateAssessment

Arguments

:: Text
$sel:name:CreateAssessment'

-> AssessmentReportsDestination
$sel:assessmentReportsDestination:CreateAssessment'

-> Scope
$sel:scope:CreateAssessment'

-> Text
$sel:frameworkId:CreateAssessment'

-> CreateAssessment

Create a value of CreateAssessment with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:description:CreateAssessment', createAssessment_description - The optional description of the assessment to be created.

$sel:tags:CreateAssessment', createAssessment_tags - The tags associated with the assessment.

$sel:name:CreateAssessment', createAssessment_name - The name of the assessment to be created.

$sel:assessmentReportsDestination:CreateAssessment', createAssessment_assessmentReportsDestination - The assessment report storage destination for the specified assessment that is being created.

$sel:scope:CreateAssessment', createAssessment_scope - Undocumented member.

$sel:roles:CreateAssessment', createAssessment_roles - The list of roles for the specified assessment.

$sel:frameworkId:CreateAssessment', createAssessment_frameworkId - The identifier for the specified framework.

Request Lenses

createAssessment_description :: Lens' CreateAssessment (Maybe Text)

The optional description of the assessment to be created.

createAssessment_tags :: Lens' CreateAssessment (Maybe (HashMap Text Text))

The tags associated with the assessment.

createAssessment_name :: Lens' CreateAssessment Text

The name of the assessment to be created.

createAssessment_assessmentReportsDestination :: Lens' CreateAssessment AssessmentReportsDestination

The assessment report storage destination for the specified assessment that is being created.

createAssessment_scope :: Lens' CreateAssessment Scope

Undocumented member.

createAssessment_roles :: Lens' CreateAssessment [Role]

The list of roles for the specified assessment.

createAssessment_frameworkId :: Lens' CreateAssessment Text

The identifier for the specified framework.

Destructuring the Response

data CreateAssessmentResponse

See: newCreateAssessmentResponse smart constructor.

Constructors

CreateAssessmentResponse'

Fields

assessment :: Maybe Assessment

httpStatus :: Int

The response's http status code.

newCreateAssessmentResponse

Arguments

:: Int
$sel:httpStatus:CreateAssessmentResponse'

-> CreateAssessmentResponse

Create a value of CreateAssessmentResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:assessment:CreateAssessmentResponse', createAssessmentResponse_assessment - Undocumented member.

$sel:httpStatus:CreateAssessmentResponse', createAssessmentResponse_httpStatus - The response's http status code.

Response Lenses

createAssessmentResponse_assessment :: Lens' CreateAssessmentResponse (Maybe Assessment)

Undocumented member.

createAssessmentResponse_httpStatus :: Lens' CreateAssessmentResponse Int

The response's http status code.