Amazonka.AuditManager.CreateAssessmentFramework

Description

Creates a custom framework in Audit Manager.

Creating a Request

data CreateAssessmentFramework

See: newCreateAssessmentFramework smart constructor.

Constructors

CreateAssessmentFramework'

Fields

complianceType :: Maybe Text

The compliance type that the new custom framework supports, such as CIS or HIPAA.

description :: Maybe Text

An optional description for the new custom framework.

tags :: Maybe (HashMap Text Text)

The tags associated with the framework.

name :: Text

The name of the new custom framework.

controlSets :: NonEmpty CreateAssessmentFrameworkControlSet

The control sets to be associated with the framework.

newCreateAssessmentFramework

Arguments

:: Text
$sel:name:CreateAssessmentFramework'

-> NonEmpty CreateAssessmentFrameworkControlSet
$sel:controlSets:CreateAssessmentFramework'

-> CreateAssessmentFramework

Create a value of CreateAssessmentFramework with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:complianceType:CreateAssessmentFramework', createAssessmentFramework_complianceType - The compliance type that the new custom framework supports, such as CIS or HIPAA.

$sel:description:CreateAssessmentFramework', createAssessmentFramework_description - An optional description for the new custom framework.

$sel:tags:CreateAssessmentFramework', createAssessmentFramework_tags - The tags associated with the framework.

$sel:name:CreateAssessmentFramework', createAssessmentFramework_name - The name of the new custom framework.

$sel:controlSets:CreateAssessmentFramework', createAssessmentFramework_controlSets - The control sets to be associated with the framework.

Request Lenses

createAssessmentFramework_complianceType :: Lens' CreateAssessmentFramework (Maybe Text)

The compliance type that the new custom framework supports, such as CIS or HIPAA.

createAssessmentFramework_description :: Lens' CreateAssessmentFramework (Maybe Text)

An optional description for the new custom framework.

createAssessmentFramework_tags :: Lens' CreateAssessmentFramework (Maybe (HashMap Text Text))

The tags associated with the framework.

createAssessmentFramework_name :: Lens' CreateAssessmentFramework Text

The name of the new custom framework.

createAssessmentFramework_controlSets :: Lens' CreateAssessmentFramework (NonEmpty CreateAssessmentFrameworkControlSet)

The control sets to be associated with the framework.

Destructuring the Response

data CreateAssessmentFrameworkResponse

See: newCreateAssessmentFrameworkResponse smart constructor.

Constructors

CreateAssessmentFrameworkResponse'

Fields

framework :: Maybe Framework

The name of the new framework returned by the CreateAssessmentFramework API.

httpStatus :: Int

The response's http status code.

newCreateAssessmentFrameworkResponse

Arguments

:: Int
$sel:httpStatus:CreateAssessmentFrameworkResponse'

-> CreateAssessmentFrameworkResponse

Create a value of CreateAssessmentFrameworkResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:framework:CreateAssessmentFrameworkResponse', createAssessmentFrameworkResponse_framework - The name of the new framework returned by the CreateAssessmentFramework API.

$sel:httpStatus:CreateAssessmentFrameworkResponse', createAssessmentFrameworkResponse_httpStatus - The response's http status code.

Response Lenses

createAssessmentFrameworkResponse_framework :: Lens' CreateAssessmentFrameworkResponse (Maybe Framework)

The name of the new framework returned by the CreateAssessmentFramework API.

createAssessmentFrameworkResponse_httpStatus :: Lens' CreateAssessmentFrameworkResponse Int

The response's http status code.