Amazonka.AuditManager.CreateControl

Description

Creates a new custom control in Audit Manager.

Creating a Request

data CreateControl

See: newCreateControl smart constructor.

Constructors

CreateControl'

Fields

testingInformation :: Maybe Text

The steps to follow to determine if the control has been satisfied.

actionPlanInstructions :: Maybe Text

The recommended actions to carry out if the control is not fulfilled.

actionPlanTitle :: Maybe Text

The title of the action plan for remediating the control.

description :: Maybe Text

The description of the control.

tags :: Maybe (HashMap Text Text)

The tags associated with the control.

name :: Text

The name of the control.

controlMappingSources :: NonEmpty CreateControlMappingSource

The data mapping sources for the specified control.

newCreateControl

Arguments

:: Text
$sel:name:CreateControl'

-> NonEmpty CreateControlMappingSource
$sel:controlMappingSources:CreateControl'

-> CreateControl

Create a value of CreateControl with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:testingInformation:CreateControl', createControl_testingInformation - The steps to follow to determine if the control has been satisfied.

$sel:actionPlanInstructions:CreateControl', createControl_actionPlanInstructions - The recommended actions to carry out if the control is not fulfilled.

$sel:actionPlanTitle:CreateControl', createControl_actionPlanTitle - The title of the action plan for remediating the control.

$sel:description:CreateControl', createControl_description - The description of the control.

$sel:tags:CreateControl', createControl_tags - The tags associated with the control.

$sel:name:CreateControl', createControl_name - The name of the control.

$sel:controlMappingSources:CreateControl', createControl_controlMappingSources - The data mapping sources for the specified control.

Request Lenses

createControl_testingInformation :: Lens' CreateControl (Maybe Text)

The steps to follow to determine if the control has been satisfied.

createControl_actionPlanInstructions :: Lens' CreateControl (Maybe Text)

The recommended actions to carry out if the control is not fulfilled.

createControl_actionPlanTitle :: Lens' CreateControl (Maybe Text)

The title of the action plan for remediating the control.

createControl_description :: Lens' CreateControl (Maybe Text)

The description of the control.

createControl_tags :: Lens' CreateControl (Maybe (HashMap Text Text))

The tags associated with the control.

createControl_name :: Lens' CreateControl Text

The name of the control.

createControl_controlMappingSources :: Lens' CreateControl (NonEmpty CreateControlMappingSource)

The data mapping sources for the specified control.

Destructuring the Response

data CreateControlResponse

See: newCreateControlResponse smart constructor.

Constructors

CreateControlResponse'

Fields

control :: Maybe Control

The new control returned by the CreateControl API.

httpStatus :: Int

The response's http status code.

newCreateControlResponse

Arguments

:: Int
$sel:httpStatus:CreateControlResponse'

-> CreateControlResponse

Create a value of CreateControlResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:control:CreateControlResponse', createControlResponse_control - The new control returned by the CreateControl API.

$sel:httpStatus:CreateControlResponse', createControlResponse_httpStatus - The response's http status code.

Response Lenses

createControlResponse_control :: Lens' CreateControlResponse (Maybe Control)

The new control returned by the CreateControl API.

createControlResponse_httpStatus :: Lens' CreateControlResponse Int

The response's http status code.