Amazonka.AuditManager.DeleteAssessment

Description

Deletes an assessment in Audit Manager.

Creating a Request

data DeleteAssessment

See: newDeleteAssessment smart constructor.

Constructors

DeleteAssessment'

Fields

assessmentId :: Text

The identifier for the specified assessment.

newDeleteAssessment

Arguments

:: Text
$sel:assessmentId:DeleteAssessment'

-> DeleteAssessment

Create a value of DeleteAssessment with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:assessmentId:DeleteAssessment', deleteAssessment_assessmentId - The identifier for the specified assessment.

Request Lenses

deleteAssessment_assessmentId :: Lens' DeleteAssessment Text

The identifier for the specified assessment.

Destructuring the Response

data DeleteAssessmentResponse

See: newDeleteAssessmentResponse smart constructor.

Constructors

DeleteAssessmentResponse'

Fields

httpStatus :: Int

The response's http status code.

newDeleteAssessmentResponse

Arguments

:: Int
$sel:httpStatus:DeleteAssessmentResponse'

-> DeleteAssessmentResponse

Create a value of DeleteAssessmentResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:httpStatus:DeleteAssessmentResponse', deleteAssessmentResponse_httpStatus - The response's http status code.

Response Lenses

deleteAssessmentResponse_httpStatus :: Lens' DeleteAssessmentResponse Int

The response's http status code.