Amazonka.AuditManager.GetAssessmentReportUrl

Description

Returns the URL of a specified assessment report in Audit Manager.

Creating a Request

data GetAssessmentReportUrl

See: newGetAssessmentReportUrl smart constructor.

Constructors

GetAssessmentReportUrl'

Fields

assessmentReportId :: Text

The identifier for the assessment report.

assessmentId :: Text

The identifier for the specified assessment.

newGetAssessmentReportUrl

Arguments

:: Text
$sel:assessmentReportId:GetAssessmentReportUrl'

-> Text
$sel:assessmentId:GetAssessmentReportUrl'

-> GetAssessmentReportUrl

Create a value of GetAssessmentReportUrl with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:assessmentReportId:GetAssessmentReportUrl', getAssessmentReportUrl_assessmentReportId - The identifier for the assessment report.

$sel:assessmentId:GetAssessmentReportUrl', getAssessmentReportUrl_assessmentId - The identifier for the specified assessment.

Request Lenses

getAssessmentReportUrl_assessmentReportId :: Lens' GetAssessmentReportUrl Text

The identifier for the assessment report.

getAssessmentReportUrl_assessmentId :: Lens' GetAssessmentReportUrl Text

The identifier for the specified assessment.

Destructuring the Response

data GetAssessmentReportUrlResponse

See: newGetAssessmentReportUrlResponse smart constructor.

Constructors

GetAssessmentReportUrlResponse'

Fields

preSignedUrl :: Maybe URL

httpStatus :: Int

The response's http status code.

newGetAssessmentReportUrlResponse

Arguments

:: Int
$sel:httpStatus:GetAssessmentReportUrlResponse'

-> GetAssessmentReportUrlResponse

Create a value of GetAssessmentReportUrlResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:preSignedUrl:GetAssessmentReportUrlResponse', getAssessmentReportUrlResponse_preSignedUrl - Undocumented member.

$sel:httpStatus:GetAssessmentReportUrlResponse', getAssessmentReportUrlResponse_httpStatus - The response's http status code.

Response Lenses

getAssessmentReportUrlResponse_preSignedUrl :: Lens' GetAssessmentReportUrlResponse (Maybe URL)

Undocumented member.

getAssessmentReportUrlResponse_httpStatus :: Lens' GetAssessmentReportUrlResponse Int

The response's http status code.