Amazonka.AuditManager.ListKeywordsForDataSource

Description

Returns a list of keywords that pre-mapped to the specified control data source.

Creating a Request

data ListKeywordsForDataSource

See: newListKeywordsForDataSource smart constructor.

Constructors

ListKeywordsForDataSource'

Fields

nextToken :: Maybe Text

The pagination token used to fetch the next set of results.

maxResults :: Maybe Natural

Represents the maximum number of results per page, or per API request call.

source :: SourceType

The control mapping data source to which the keywords apply.

newListKeywordsForDataSource

Arguments

:: SourceType
$sel:source:ListKeywordsForDataSource'

-> ListKeywordsForDataSource

Create a value of ListKeywordsForDataSource with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:nextToken:ListKeywordsForDataSource', listKeywordsForDataSource_nextToken - The pagination token used to fetch the next set of results.

$sel:maxResults:ListKeywordsForDataSource', listKeywordsForDataSource_maxResults - Represents the maximum number of results per page, or per API request call.

$sel:source:ListKeywordsForDataSource', listKeywordsForDataSource_source - The control mapping data source to which the keywords apply.

Request Lenses

listKeywordsForDataSource_nextToken :: Lens' ListKeywordsForDataSource (Maybe Text)

The pagination token used to fetch the next set of results.

listKeywordsForDataSource_maxResults :: Lens' ListKeywordsForDataSource (Maybe Natural)

Represents the maximum number of results per page, or per API request call.

listKeywordsForDataSource_source :: Lens' ListKeywordsForDataSource SourceType

The control mapping data source to which the keywords apply.

Destructuring the Response

data ListKeywordsForDataSourceResponse

See: newListKeywordsForDataSourceResponse smart constructor.

Constructors

ListKeywordsForDataSourceResponse'

Fields

nextToken :: Maybe Text

The pagination token used to fetch the next set of results.

keywords :: Maybe [Text]

The list of keywords for the specified event mapping source.

httpStatus :: Int

The response's http status code.

newListKeywordsForDataSourceResponse

Arguments

:: Int
$sel:httpStatus:ListKeywordsForDataSourceResponse'

-> ListKeywordsForDataSourceResponse

Create a value of ListKeywordsForDataSourceResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:nextToken:ListKeywordsForDataSource', listKeywordsForDataSourceResponse_nextToken - The pagination token used to fetch the next set of results.

$sel:keywords:ListKeywordsForDataSourceResponse', listKeywordsForDataSourceResponse_keywords - The list of keywords for the specified event mapping source.

$sel:httpStatus:ListKeywordsForDataSourceResponse', listKeywordsForDataSourceResponse_httpStatus - The response's http status code.

Response Lenses

listKeywordsForDataSourceResponse_nextToken :: Lens' ListKeywordsForDataSourceResponse (Maybe Text)

The pagination token used to fetch the next set of results.

listKeywordsForDataSourceResponse_keywords :: Lens' ListKeywordsForDataSourceResponse (Maybe [Text])

The list of keywords for the specified event mapping source.

listKeywordsForDataSourceResponse_httpStatus :: Lens' ListKeywordsForDataSourceResponse Int

The response's http status code.