Amazonka.AuditManager.Types.AssessmentControlSet

Description

Documentation

data AssessmentControlSet

Represents a set of controls in an Audit Manager assessment.

See: newAssessmentControlSet smart constructor.

Constructors

AssessmentControlSet'

Fields

status :: Maybe ControlSetStatus

Specifies the current status of the control set.

controls :: Maybe [AssessmentControl]

The list of controls contained with the control set.

roles :: Maybe [Role]

The roles associated with the control set.

manualEvidenceCount :: Maybe Int

The total number of evidence objects uploaded manually to the control set.

delegations :: Maybe [Delegation]

The delegations associated with the control set.

systemEvidenceCount :: Maybe Int

The total number of evidence objects retrieved automatically for the control set.

id :: Maybe Text

The identifier of the control set in the assessment. This is the control set name in a plain string format.

description :: Maybe Text

The description for the control set.

newAssessmentControlSet :: AssessmentControlSet

Create a value of AssessmentControlSet with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:status:AssessmentControlSet', assessmentControlSet_status - Specifies the current status of the control set.

$sel:controls:AssessmentControlSet', assessmentControlSet_controls - The list of controls contained with the control set.

$sel:roles:AssessmentControlSet', assessmentControlSet_roles - The roles associated with the control set.

$sel:manualEvidenceCount:AssessmentControlSet', assessmentControlSet_manualEvidenceCount - The total number of evidence objects uploaded manually to the control set.

$sel:delegations:AssessmentControlSet', assessmentControlSet_delegations - The delegations associated with the control set.

$sel:systemEvidenceCount:AssessmentControlSet', assessmentControlSet_systemEvidenceCount - The total number of evidence objects retrieved automatically for the control set.

$sel:id:AssessmentControlSet', assessmentControlSet_id - The identifier of the control set in the assessment. This is the control set name in a plain string format.

$sel:description:AssessmentControlSet', assessmentControlSet_description - The description for the control set.

assessmentControlSet_status :: Lens' AssessmentControlSet (Maybe ControlSetStatus)

Specifies the current status of the control set.

assessmentControlSet_controls :: Lens' AssessmentControlSet (Maybe [AssessmentControl])

The list of controls contained with the control set.

assessmentControlSet_roles :: Lens' AssessmentControlSet (Maybe [Role])

The roles associated with the control set.

assessmentControlSet_manualEvidenceCount :: Lens' AssessmentControlSet (Maybe Int)

The total number of evidence objects uploaded manually to the control set.

assessmentControlSet_delegations :: Lens' AssessmentControlSet (Maybe [Delegation])

The delegations associated with the control set.

assessmentControlSet_systemEvidenceCount :: Lens' AssessmentControlSet (Maybe Int)

The total number of evidence objects retrieved automatically for the control set.

assessmentControlSet_id :: Lens' AssessmentControlSet (Maybe Text)

The identifier of the control set in the assessment. This is the control set name in a plain string format.

assessmentControlSet_description :: Lens' AssessmentControlSet (Maybe Text)

The description for the control set.