Amazonka.AuditManager.Types.AssessmentMetadata

Description

Documentation

data AssessmentMetadata

The metadata associated with the specified assessment.

See: newAssessmentMetadata smart constructor.

Constructors

AssessmentMetadata'

Fields

creationTime :: Maybe POSIX

Specifies when the assessment was created.

status :: Maybe AssessmentStatus

The overall status of the assessment.

lastUpdated :: Maybe POSIX

The time of the most recent update.

roles :: Maybe [Role]

The roles associated with the assessment.

delegations :: Maybe [Delegation]

The delegations associated with the assessment.

name :: Maybe Text

The name of the assessment.

assessmentReportsDestination :: Maybe AssessmentReportsDestination

The destination in which evidence reports are stored for the specified assessment.

scope :: Maybe Scope

The wrapper of Amazon Web Services accounts and services in scope for the assessment.

complianceType :: Maybe Text

The name of a compliance standard related to the assessment, such as PCI-DSS.

id :: Maybe Text

The unique identifier for the assessment.

description :: Maybe Text

The description of the assessment.

newAssessmentMetadata :: AssessmentMetadata

Create a value of AssessmentMetadata with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:creationTime:AssessmentMetadata', assessmentMetadata_creationTime - Specifies when the assessment was created.

$sel:status:AssessmentMetadata', assessmentMetadata_status - The overall status of the assessment.

$sel:lastUpdated:AssessmentMetadata', assessmentMetadata_lastUpdated - The time of the most recent update.

$sel:roles:AssessmentMetadata', assessmentMetadata_roles - The roles associated with the assessment.

$sel:delegations:AssessmentMetadata', assessmentMetadata_delegations - The delegations associated with the assessment.

$sel:name:AssessmentMetadata', assessmentMetadata_name - The name of the assessment.

$sel:assessmentReportsDestination:AssessmentMetadata', assessmentMetadata_assessmentReportsDestination - The destination in which evidence reports are stored for the specified assessment.

$sel:scope:AssessmentMetadata', assessmentMetadata_scope - The wrapper of Amazon Web Services accounts and services in scope for the assessment.

$sel:complianceType:AssessmentMetadata', assessmentMetadata_complianceType - The name of a compliance standard related to the assessment, such as PCI-DSS.

$sel:id:AssessmentMetadata', assessmentMetadata_id - The unique identifier for the assessment.

$sel:description:AssessmentMetadata', assessmentMetadata_description - The description of the assessment.

assessmentMetadata_creationTime :: Lens' AssessmentMetadata (Maybe UTCTime)

Specifies when the assessment was created.

assessmentMetadata_status :: Lens' AssessmentMetadata (Maybe AssessmentStatus)

The overall status of the assessment.

assessmentMetadata_lastUpdated :: Lens' AssessmentMetadata (Maybe UTCTime)

The time of the most recent update.

assessmentMetadata_roles :: Lens' AssessmentMetadata (Maybe [Role])

The roles associated with the assessment.

assessmentMetadata_delegations :: Lens' AssessmentMetadata (Maybe [Delegation])

The delegations associated with the assessment.

assessmentMetadata_name :: Lens' AssessmentMetadata (Maybe Text)

The name of the assessment.

assessmentMetadata_assessmentReportsDestination :: Lens' AssessmentMetadata (Maybe AssessmentReportsDestination)

The destination in which evidence reports are stored for the specified assessment.

assessmentMetadata_scope :: Lens' AssessmentMetadata (Maybe Scope)

The wrapper of Amazon Web Services accounts and services in scope for the assessment.

assessmentMetadata_complianceType :: Lens' AssessmentMetadata (Maybe Text)

The name of a compliance standard related to the assessment, such as PCI-DSS.

assessmentMetadata_id :: Lens' AssessmentMetadata (Maybe Text)

The unique identifier for the assessment.

assessmentMetadata_description :: Lens' AssessmentMetadata (Maybe Text)

The description of the assessment.