Amazonka.AuditManager.Types.AssessmentMetadataItem

Description

Documentation

data AssessmentMetadataItem

A metadata object associated with an assessment in Audit Manager.

See: newAssessmentMetadataItem smart constructor.

Constructors

AssessmentMetadataItem'

Fields

creationTime :: Maybe POSIX

Specifies when the assessment was created.

status :: Maybe AssessmentStatus

The current status of the assessment.

lastUpdated :: Maybe POSIX

The time of the most recent update.

roles :: Maybe [Role]

The roles associated with the assessment.

delegations :: Maybe [Delegation]

The delegations associated with the assessment.

name :: Maybe Text

The name of the assessment.

complianceType :: Maybe Text

The name of the compliance standard related to the assessment, such as PCI-DSS.

id :: Maybe Text

The unique identifier for the assessment.

newAssessmentMetadataItem :: AssessmentMetadataItem

Create a value of AssessmentMetadataItem with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:creationTime:AssessmentMetadataItem', assessmentMetadataItem_creationTime - Specifies when the assessment was created.

$sel:status:AssessmentMetadataItem', assessmentMetadataItem_status - The current status of the assessment.

$sel:lastUpdated:AssessmentMetadataItem', assessmentMetadataItem_lastUpdated - The time of the most recent update.

$sel:roles:AssessmentMetadataItem', assessmentMetadataItem_roles - The roles associated with the assessment.

$sel:delegations:AssessmentMetadataItem', assessmentMetadataItem_delegations - The delegations associated with the assessment.

$sel:name:AssessmentMetadataItem', assessmentMetadataItem_name - The name of the assessment.

$sel:complianceType:AssessmentMetadataItem', assessmentMetadataItem_complianceType - The name of the compliance standard related to the assessment, such as PCI-DSS.

$sel:id:AssessmentMetadataItem', assessmentMetadataItem_id - The unique identifier for the assessment.

assessmentMetadataItem_creationTime :: Lens' AssessmentMetadataItem (Maybe UTCTime)

Specifies when the assessment was created.

assessmentMetadataItem_status :: Lens' AssessmentMetadataItem (Maybe AssessmentStatus)

The current status of the assessment.

assessmentMetadataItem_lastUpdated :: Lens' AssessmentMetadataItem (Maybe UTCTime)

The time of the most recent update.

assessmentMetadataItem_roles :: Lens' AssessmentMetadataItem (Maybe [Role])

The roles associated with the assessment.

assessmentMetadataItem_delegations :: Lens' AssessmentMetadataItem (Maybe [Delegation])

The delegations associated with the assessment.

assessmentMetadataItem_name :: Lens' AssessmentMetadataItem (Maybe Text)

The name of the assessment.

assessmentMetadataItem_complianceType :: Lens' AssessmentMetadataItem (Maybe Text)

The name of the compliance standard related to the assessment, such as PCI-DSS.

assessmentMetadataItem_id :: Lens' AssessmentMetadataItem (Maybe Text)

The unique identifier for the assessment.