Amazonka.AuditManager.Types.Control

Description

Documentation

data Control

A control in Audit Manager.

See: newControl smart constructor.

Constructors

Control'

Fields

lastUpdatedBy :: Maybe Text

The IAM user or role that most recently updated the control.

testingInformation :: Maybe Text

The steps to follow to determine if the control has been satisfied.

lastUpdatedAt :: Maybe POSIX

Specifies when the control was most recently updated.

arn :: Maybe Text

The Amazon Resource Name (ARN) of the specified control.

createdAt :: Maybe POSIX

Specifies when the control was created.

controlMappingSources :: Maybe (NonEmpty ControlMappingSource)

The data mapping sources for the specified control.

createdBy :: Maybe Text

The IAM user or role that created the control.

actionPlanInstructions :: Maybe Text

The recommended actions to carry out if the control is not fulfilled.

controlSources :: Maybe Text

The data source that determines from where Audit Manager collects evidence for the control.

name :: Maybe Text

The name of the specified control.

actionPlanTitle :: Maybe Text

The title of the action plan for remediating the control.

id :: Maybe Text

The unique identifier for the control.

type' :: Maybe ControlType

The type of control, such as custom or standard.

description :: Maybe Text

The description of the specified control.

tags :: Maybe (HashMap Text Text)

The tags associated with the control.

newControl :: Control

Create a value of Control with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:lastUpdatedBy:Control', control_lastUpdatedBy - The IAM user or role that most recently updated the control.

$sel:testingInformation:Control', control_testingInformation - The steps to follow to determine if the control has been satisfied.

$sel:lastUpdatedAt:Control', control_lastUpdatedAt - Specifies when the control was most recently updated.

$sel:arn:Control', control_arn - The Amazon Resource Name (ARN) of the specified control.

$sel:createdAt:Control', control_createdAt - Specifies when the control was created.

$sel:controlMappingSources:Control', control_controlMappingSources - The data mapping sources for the specified control.

$sel:createdBy:Control', control_createdBy - The IAM user or role that created the control.

$sel:actionPlanInstructions:Control', control_actionPlanInstructions - The recommended actions to carry out if the control is not fulfilled.

$sel:controlSources:Control', control_controlSources - The data source that determines from where Audit Manager collects evidence for the control.

$sel:name:Control', control_name - The name of the specified control.

$sel:actionPlanTitle:Control', control_actionPlanTitle - The title of the action plan for remediating the control.

$sel:id:Control', control_id - The unique identifier for the control.

$sel:type':Control', control_type - The type of control, such as custom or standard.

$sel:description:Control', control_description - The description of the specified control.

$sel:tags:Control', control_tags - The tags associated with the control.

control_lastUpdatedBy :: Lens' Control (Maybe Text)

The IAM user or role that most recently updated the control.

control_testingInformation :: Lens' Control (Maybe Text)

The steps to follow to determine if the control has been satisfied.

control_lastUpdatedAt :: Lens' Control (Maybe UTCTime)

Specifies when the control was most recently updated.

control_arn :: Lens' Control (Maybe Text)

The Amazon Resource Name (ARN) of the specified control.

control_createdAt :: Lens' Control (Maybe UTCTime)

Specifies when the control was created.

control_controlMappingSources :: Lens' Control (Maybe (NonEmpty ControlMappingSource))

The data mapping sources for the specified control.

control_createdBy :: Lens' Control (Maybe Text)

The IAM user or role that created the control.

control_actionPlanInstructions :: Lens' Control (Maybe Text)

The recommended actions to carry out if the control is not fulfilled.

control_controlSources :: Lens' Control (Maybe Text)

The data source that determines from where Audit Manager collects evidence for the control.

control_name :: Lens' Control (Maybe Text)

The name of the specified control.

control_actionPlanTitle :: Lens' Control (Maybe Text)

The title of the action plan for remediating the control.

control_id :: Lens' Control (Maybe Text)

The unique identifier for the control.

control_type :: Lens' Control (Maybe ControlType)

The type of control, such as custom or standard.

control_description :: Lens' Control (Maybe Text)

The description of the specified control.

control_tags :: Lens' Control (Maybe (HashMap Text Text))

The tags associated with the control.