Amazonka.AuditManager.Types.ControlMappingSource

Description

Documentation

data ControlMappingSource

The data source that determines from where Audit Manager collects evidence for the control.

See: newControlMappingSource smart constructor.

Constructors

ControlMappingSource'

Fields

sourceName :: Maybe Text

The name of the specified source.

sourceType :: Maybe SourceType

Specifies one of the five types of data sources for evidence collection.

troubleshootingText :: Maybe Text

The instructions for troubleshooting the specified control.

sourceId :: Maybe Text

The unique identifier for the specified source.

sourceDescription :: Maybe Text

The description of the specified source.

sourceFrequency :: Maybe SourceFrequency

The frequency of evidence collection for the specified control mapping source.

sourceKeyword :: Maybe SourceKeyword

sourceSetUpOption :: Maybe SourceSetUpOption

The setup option for the data source, which reflects if the evidence collection is automated or manual.

newControlMappingSource :: ControlMappingSource

Create a value of ControlMappingSource with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:sourceName:ControlMappingSource', controlMappingSource_sourceName - The name of the specified source.

$sel:sourceType:ControlMappingSource', controlMappingSource_sourceType - Specifies one of the five types of data sources for evidence collection.

$sel:troubleshootingText:ControlMappingSource', controlMappingSource_troubleshootingText - The instructions for troubleshooting the specified control.

$sel:sourceId:ControlMappingSource', controlMappingSource_sourceId - The unique identifier for the specified source.

$sel:sourceDescription:ControlMappingSource', controlMappingSource_sourceDescription - The description of the specified source.

$sel:sourceFrequency:ControlMappingSource', controlMappingSource_sourceFrequency - The frequency of evidence collection for the specified control mapping source.

$sel:sourceKeyword:ControlMappingSource', controlMappingSource_sourceKeyword - Undocumented member.

$sel:sourceSetUpOption:ControlMappingSource', controlMappingSource_sourceSetUpOption - The setup option for the data source, which reflects if the evidence collection is automated or manual.

controlMappingSource_sourceName :: Lens' ControlMappingSource (Maybe Text)

The name of the specified source.

controlMappingSource_sourceType :: Lens' ControlMappingSource (Maybe SourceType)

Specifies one of the five types of data sources for evidence collection.

controlMappingSource_troubleshootingText :: Lens' ControlMappingSource (Maybe Text)

The instructions for troubleshooting the specified control.

controlMappingSource_sourceId :: Lens' ControlMappingSource (Maybe Text)

The unique identifier for the specified source.

controlMappingSource_sourceDescription :: Lens' ControlMappingSource (Maybe Text)

The description of the specified source.

controlMappingSource_sourceFrequency :: Lens' ControlMappingSource (Maybe SourceFrequency)

The frequency of evidence collection for the specified control mapping source.

controlMappingSource_sourceKeyword :: Lens' ControlMappingSource (Maybe SourceKeyword)

Undocumented member.

controlMappingSource_sourceSetUpOption :: Lens' ControlMappingSource (Maybe SourceSetUpOption)

The setup option for the data source, which reflects if the evidence collection is automated or manual.