Amazonka.AuditManager.Types.CreateControlMappingSource

Description

Documentation

data CreateControlMappingSource

Control mapping fields that represent the source for evidence collection, along with related parameters and metadata. This does not contain mappingID.

See: newCreateControlMappingSource smart constructor.

Constructors

CreateControlMappingSource'

Fields

sourceName :: Maybe Text

The name of the control mapping data source.

sourceType :: Maybe SourceType

Specifies one of the five types of data sources for evidence collection.

troubleshootingText :: Maybe Text

The instructions for troubleshooting the specified control.

sourceDescription :: Maybe Text

The description of the data source that determines from where Audit Manager collects evidence for the control.

sourceFrequency :: Maybe SourceFrequency

The frequency of evidence collection for the specified control mapping source.

sourceKeyword :: Maybe SourceKeyword

sourceSetUpOption :: Maybe SourceSetUpOption

The setup option for the data source, which reflects if the evidence collection is automated or manual.

newCreateControlMappingSource :: CreateControlMappingSource

Create a value of CreateControlMappingSource with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:sourceName:CreateControlMappingSource', createControlMappingSource_sourceName - The name of the control mapping data source.

$sel:sourceType:CreateControlMappingSource', createControlMappingSource_sourceType - Specifies one of the five types of data sources for evidence collection.

$sel:troubleshootingText:CreateControlMappingSource', createControlMappingSource_troubleshootingText - The instructions for troubleshooting the specified control.

$sel:sourceDescription:CreateControlMappingSource', createControlMappingSource_sourceDescription - The description of the data source that determines from where Audit Manager collects evidence for the control.

$sel:sourceFrequency:CreateControlMappingSource', createControlMappingSource_sourceFrequency - The frequency of evidence collection for the specified control mapping source.

$sel:sourceKeyword:CreateControlMappingSource', createControlMappingSource_sourceKeyword - Undocumented member.

$sel:sourceSetUpOption:CreateControlMappingSource', createControlMappingSource_sourceSetUpOption - The setup option for the data source, which reflects if the evidence collection is automated or manual.

createControlMappingSource_sourceName :: Lens' CreateControlMappingSource (Maybe Text)

The name of the control mapping data source.

createControlMappingSource_sourceType :: Lens' CreateControlMappingSource (Maybe SourceType)

Specifies one of the five types of data sources for evidence collection.

createControlMappingSource_troubleshootingText :: Lens' CreateControlMappingSource (Maybe Text)

The instructions for troubleshooting the specified control.

createControlMappingSource_sourceDescription :: Lens' CreateControlMappingSource (Maybe Text)

The description of the data source that determines from where Audit Manager collects evidence for the control.

createControlMappingSource_sourceFrequency :: Lens' CreateControlMappingSource (Maybe SourceFrequency)

The frequency of evidence collection for the specified control mapping source.

createControlMappingSource_sourceKeyword :: Lens' CreateControlMappingSource (Maybe SourceKeyword)

Undocumented member.

createControlMappingSource_sourceSetUpOption :: Lens' CreateControlMappingSource (Maybe SourceSetUpOption)

The setup option for the data source, which reflects if the evidence collection is automated or manual.