Amazonka.AuditManager.Types.Framework

Description

Documentation

data Framework

The file used to structure and automate Audit Manager assessments for a given compliance standard.

See: newFramework smart constructor.

Constructors

Framework'

Fields

lastUpdatedBy :: Maybe Text

The IAM user or role that most recently updated the framework.

lastUpdatedAt :: Maybe POSIX

Specifies when the framework was most recently updated.

arn :: Maybe Text

The Amazon Resource Name (ARN) of the specified framework.

createdAt :: Maybe POSIX

Specifies when the framework was created.

createdBy :: Maybe Text

The IAM user or role that created the framework.

controlSets :: Maybe (NonEmpty ControlSet)

The control sets associated with the framework.

controlSources :: Maybe Text

The sources from which Audit Manager collects evidence for the control.

name :: Maybe Text

The name of the specified framework.

complianceType :: Maybe Text

The compliance type that the new custom framework supports, such as CIS or HIPAA.

id :: Maybe Text

The unique identifier for the specified framework.

type' :: Maybe FrameworkType

The framework type, such as custom or standard.

logo :: Maybe Text

The logo associated with the framework.

description :: Maybe Text

The description of the specified framework.

tags :: Maybe (HashMap Text Text)

The tags associated with the framework.

newFramework :: Framework

Create a value of Framework with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:lastUpdatedBy:Framework', framework_lastUpdatedBy - The IAM user or role that most recently updated the framework.

$sel:lastUpdatedAt:Framework', framework_lastUpdatedAt - Specifies when the framework was most recently updated.

$sel:arn:Framework', framework_arn - The Amazon Resource Name (ARN) of the specified framework.

$sel:createdAt:Framework', framework_createdAt - Specifies when the framework was created.

$sel:createdBy:Framework', framework_createdBy - The IAM user or role that created the framework.

$sel:controlSets:Framework', framework_controlSets - The control sets associated with the framework.

$sel:controlSources:Framework', framework_controlSources - The sources from which Audit Manager collects evidence for the control.

$sel:name:Framework', framework_name - The name of the specified framework.

$sel:complianceType:Framework', framework_complianceType - The compliance type that the new custom framework supports, such as CIS or HIPAA.

$sel:id:Framework', framework_id - The unique identifier for the specified framework.

$sel:type':Framework', framework_type - The framework type, such as custom or standard.

$sel:logo:Framework', framework_logo - The logo associated with the framework.

$sel:description:Framework', framework_description - The description of the specified framework.

$sel:tags:Framework', framework_tags - The tags associated with the framework.

framework_lastUpdatedBy :: Lens' Framework (Maybe Text)

The IAM user or role that most recently updated the framework.

framework_lastUpdatedAt :: Lens' Framework (Maybe UTCTime)

Specifies when the framework was most recently updated.

framework_arn :: Lens' Framework (Maybe Text)

The Amazon Resource Name (ARN) of the specified framework.

framework_createdAt :: Lens' Framework (Maybe UTCTime)

Specifies when the framework was created.

framework_createdBy :: Lens' Framework (Maybe Text)

The IAM user or role that created the framework.

framework_controlSets :: Lens' Framework (Maybe (NonEmpty ControlSet))

The control sets associated with the framework.

framework_controlSources :: Lens' Framework (Maybe Text)

The sources from which Audit Manager collects evidence for the control.

framework_name :: Lens' Framework (Maybe Text)

The name of the specified framework.

framework_complianceType :: Lens' Framework (Maybe Text)

The compliance type that the new custom framework supports, such as CIS or HIPAA.

framework_id :: Lens' Framework (Maybe Text)

The unique identifier for the specified framework.

framework_type :: Lens' Framework (Maybe FrameworkType)

The framework type, such as custom or standard.

framework_logo :: Lens' Framework (Maybe Text)

The logo associated with the framework.

framework_description :: Lens' Framework (Maybe Text)

The description of the specified framework.

framework_tags :: Lens' Framework (Maybe (HashMap Text Text))

The tags associated with the framework.