Amazonka.AuditManager.Types.Settings

Description

Documentation

data Settings

The settings object that holds all supported Audit Manager settings.

See: newSettings smart constructor.

Constructors

Settings'

Fields

kmsKey :: Maybe Text

The KMS key details.

defaultAssessmentReportsDestination :: Maybe AssessmentReportsDestination

The default storage destination for assessment reports.

snsTopic :: Maybe Text

The designated Amazon Simple Notification Service (Amazon SNS) topic.

defaultProcessOwners :: Maybe [Role]

The designated default audit owners.

isAwsOrgEnabled :: Maybe Bool

Specifies whether Organizations is enabled.

newSettings :: Settings

Create a value of Settings with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:kmsKey:Settings', settings_kmsKey - The KMS key details.

$sel:defaultAssessmentReportsDestination:Settings', settings_defaultAssessmentReportsDestination - The default storage destination for assessment reports.

$sel:snsTopic:Settings', settings_snsTopic - The designated Amazon Simple Notification Service (Amazon SNS) topic.

$sel:defaultProcessOwners:Settings', settings_defaultProcessOwners - The designated default audit owners.

$sel:isAwsOrgEnabled:Settings', settings_isAwsOrgEnabled - Specifies whether Organizations is enabled.

settings_kmsKey :: Lens' Settings (Maybe Text)

The KMS key details.

settings_defaultAssessmentReportsDestination :: Lens' Settings (Maybe AssessmentReportsDestination)

The default storage destination for assessment reports.

settings_snsTopic :: Lens' Settings (Maybe Text)

The designated Amazon Simple Notification Service (Amazon SNS) topic.

settings_defaultProcessOwners :: Lens' Settings (Maybe [Role])

The designated default audit owners.

settings_isAwsOrgEnabled :: Lens' Settings (Maybe Bool)

Specifies whether Organizations is enabled.