Amazonka.AuditManager.UpdateAssessment

Description

Edits an Audit Manager assessment.

Creating a Request

data UpdateAssessment

See: newUpdateAssessment smart constructor.

Constructors

UpdateAssessment'

Fields

roles :: Maybe [Role]

The list of roles for the specified assessment.

assessmentDescription :: Maybe Text

The description of the specified assessment.

assessmentReportsDestination :: Maybe AssessmentReportsDestination

The assessment report storage destination for the specified assessment that is being updated.

assessmentName :: Maybe Text

The name of the specified assessment to be updated.

assessmentId :: Text

The identifier for the specified assessment.

scope :: Scope

The scope of the specified assessment.

newUpdateAssessment

Arguments

:: Text
$sel:assessmentId:UpdateAssessment'

-> Scope
$sel:scope:UpdateAssessment'

-> UpdateAssessment

Create a value of UpdateAssessment with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:roles:UpdateAssessment', updateAssessment_roles - The list of roles for the specified assessment.

$sel:assessmentDescription:UpdateAssessment', updateAssessment_assessmentDescription - The description of the specified assessment.

$sel:assessmentReportsDestination:UpdateAssessment', updateAssessment_assessmentReportsDestination - The assessment report storage destination for the specified assessment that is being updated.

$sel:assessmentName:UpdateAssessment', updateAssessment_assessmentName - The name of the specified assessment to be updated.

$sel:assessmentId:UpdateAssessment', updateAssessment_assessmentId - The identifier for the specified assessment.

$sel:scope:UpdateAssessment', updateAssessment_scope - The scope of the specified assessment.

Request Lenses

updateAssessment_roles :: Lens' UpdateAssessment (Maybe [Role])

The list of roles for the specified assessment.

updateAssessment_assessmentDescription :: Lens' UpdateAssessment (Maybe Text)

The description of the specified assessment.

updateAssessment_assessmentReportsDestination :: Lens' UpdateAssessment (Maybe AssessmentReportsDestination)

The assessment report storage destination for the specified assessment that is being updated.

updateAssessment_assessmentName :: Lens' UpdateAssessment (Maybe Text)

The name of the specified assessment to be updated.

updateAssessment_assessmentId :: Lens' UpdateAssessment Text

The identifier for the specified assessment.

updateAssessment_scope :: Lens' UpdateAssessment Scope

The scope of the specified assessment.

Destructuring the Response

data UpdateAssessmentResponse

See: newUpdateAssessmentResponse smart constructor.

Constructors

UpdateAssessmentResponse'

Fields

assessment :: Maybe Assessment

The response object (name of the updated assessment) for the UpdateAssessmentRequest API.

httpStatus :: Int

The response's http status code.

newUpdateAssessmentResponse

Arguments

:: Int
$sel:httpStatus:UpdateAssessmentResponse'

-> UpdateAssessmentResponse

Create a value of UpdateAssessmentResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:assessment:UpdateAssessmentResponse', updateAssessmentResponse_assessment - The response object (name of the updated assessment) for the UpdateAssessmentRequest API.

$sel:httpStatus:UpdateAssessmentResponse', updateAssessmentResponse_httpStatus - The response's http status code.

Response Lenses

updateAssessmentResponse_assessment :: Lens' UpdateAssessmentResponse (Maybe Assessment)

The response object (name of the updated assessment) for the UpdateAssessmentRequest API.

updateAssessmentResponse_httpStatus :: Lens' UpdateAssessmentResponse Int

The response's http status code.