Amazonka.AuditManager.UpdateAssessmentControl

Description

Updates a control within an assessment in Audit Manager.

Creating a Request

data UpdateAssessmentControl

See: newUpdateAssessmentControl smart constructor.

Constructors

UpdateAssessmentControl'

Fields

controlStatus :: Maybe ControlStatus

The status of the specified control.

commentBody :: Maybe Text

The comment body text for the specified control.

assessmentId :: Text

The identifier for the specified assessment.

controlSetId :: Text

The identifier for the specified control set.

controlId :: Text

The identifier for the specified control.

newUpdateAssessmentControl

Arguments

:: Text
$sel:assessmentId:UpdateAssessmentControl'

-> Text
$sel:controlSetId:UpdateAssessmentControl'

-> Text
$sel:controlId:UpdateAssessmentControl'

-> UpdateAssessmentControl

Create a value of UpdateAssessmentControl with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:controlStatus:UpdateAssessmentControl', updateAssessmentControl_controlStatus - The status of the specified control.

$sel:commentBody:UpdateAssessmentControl', updateAssessmentControl_commentBody - The comment body text for the specified control.

$sel:assessmentId:UpdateAssessmentControl', updateAssessmentControl_assessmentId - The identifier for the specified assessment.

$sel:controlSetId:UpdateAssessmentControl', updateAssessmentControl_controlSetId - The identifier for the specified control set.

$sel:controlId:UpdateAssessmentControl', updateAssessmentControl_controlId - The identifier for the specified control.

Request Lenses

updateAssessmentControl_controlStatus :: Lens' UpdateAssessmentControl (Maybe ControlStatus)

The status of the specified control.

updateAssessmentControl_commentBody :: Lens' UpdateAssessmentControl (Maybe Text)

The comment body text for the specified control.

updateAssessmentControl_assessmentId :: Lens' UpdateAssessmentControl Text

The identifier for the specified assessment.

updateAssessmentControl_controlSetId :: Lens' UpdateAssessmentControl Text

The identifier for the specified control set.

updateAssessmentControl_controlId :: Lens' UpdateAssessmentControl Text

The identifier for the specified control.

Destructuring the Response

data UpdateAssessmentControlResponse

See: newUpdateAssessmentControlResponse smart constructor.

Constructors

UpdateAssessmentControlResponse'

Fields

control :: Maybe AssessmentControl

The name of the updated control set returned by the UpdateAssessmentControl API.

httpStatus :: Int

The response's http status code.

newUpdateAssessmentControlResponse

Arguments

:: Int
$sel:httpStatus:UpdateAssessmentControlResponse'

-> UpdateAssessmentControlResponse

Create a value of UpdateAssessmentControlResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:control:UpdateAssessmentControlResponse', updateAssessmentControlResponse_control - The name of the updated control set returned by the UpdateAssessmentControl API.

$sel:httpStatus:UpdateAssessmentControlResponse', updateAssessmentControlResponse_httpStatus - The response's http status code.

Response Lenses

updateAssessmentControlResponse_control :: Lens' UpdateAssessmentControlResponse (Maybe AssessmentControl)

The name of the updated control set returned by the UpdateAssessmentControl API.

updateAssessmentControlResponse_httpStatus :: Lens' UpdateAssessmentControlResponse Int

The response's http status code.