Amazonka.AuditManager.ValidateAssessmentReportIntegrity

Description

Validates the integrity of an assessment report in Audit Manager.

Creating a Request

data ValidateAssessmentReportIntegrity

See: newValidateAssessmentReportIntegrity smart constructor.

Constructors

ValidateAssessmentReportIntegrity'

Fields

s3RelativePath :: Text

The relative path of the specified Amazon S3 bucket in which the assessment report is stored.

newValidateAssessmentReportIntegrity

Arguments

:: Text
$sel:s3RelativePath:ValidateAssessmentReportIntegrity'

-> ValidateAssessmentReportIntegrity

Create a value of ValidateAssessmentReportIntegrity with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:s3RelativePath:ValidateAssessmentReportIntegrity', validateAssessmentReportIntegrity_s3RelativePath - The relative path of the specified Amazon S3 bucket in which the assessment report is stored.

Request Lenses

validateAssessmentReportIntegrity_s3RelativePath :: Lens' ValidateAssessmentReportIntegrity Text

The relative path of the specified Amazon S3 bucket in which the assessment report is stored.

Destructuring the Response

data ValidateAssessmentReportIntegrityResponse

See: newValidateAssessmentReportIntegrityResponse smart constructor.

Constructors

ValidateAssessmentReportIntegrityResponse'

Fields

signatureValid :: Maybe Bool

Specifies whether the signature key is valid.

validationErrors :: Maybe [Text]

Represents any errors that occurred when validating the assessment report.

signatureDateTime :: Maybe Text

The date and time signature that specifies when the assessment report was created.

signatureAlgorithm :: Maybe Text

The signature algorithm used to code sign the assessment report file.

signatureKeyId :: Maybe Text

The unique identifier for the validation signature key.

httpStatus :: Int

The response's http status code.

newValidateAssessmentReportIntegrityResponse

Arguments

:: Int
$sel:httpStatus:ValidateAssessmentReportIntegrityResponse'

-> ValidateAssessmentReportIntegrityResponse

Create a value of ValidateAssessmentReportIntegrityResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:signatureValid:ValidateAssessmentReportIntegrityResponse', validateAssessmentReportIntegrityResponse_signatureValid - Specifies whether the signature key is valid.

$sel:validationErrors:ValidateAssessmentReportIntegrityResponse', validateAssessmentReportIntegrityResponse_validationErrors - Represents any errors that occurred when validating the assessment report.

$sel:signatureDateTime:ValidateAssessmentReportIntegrityResponse', validateAssessmentReportIntegrityResponse_signatureDateTime - The date and time signature that specifies when the assessment report was created.

$sel:signatureAlgorithm:ValidateAssessmentReportIntegrityResponse', validateAssessmentReportIntegrityResponse_signatureAlgorithm - The signature algorithm used to code sign the assessment report file.

$sel:signatureKeyId:ValidateAssessmentReportIntegrityResponse', validateAssessmentReportIntegrityResponse_signatureKeyId - The unique identifier for the validation signature key.

$sel:httpStatus:ValidateAssessmentReportIntegrityResponse', validateAssessmentReportIntegrityResponse_httpStatus - The response's http status code.

Response Lenses

validateAssessmentReportIntegrityResponse_signatureValid :: Lens' ValidateAssessmentReportIntegrityResponse (Maybe Bool)

Specifies whether the signature key is valid.

validateAssessmentReportIntegrityResponse_validationErrors :: Lens' ValidateAssessmentReportIntegrityResponse (Maybe [Text])

Represents any errors that occurred when validating the assessment report.

validateAssessmentReportIntegrityResponse_signatureDateTime :: Lens' ValidateAssessmentReportIntegrityResponse (Maybe Text)

The date and time signature that specifies when the assessment report was created.

validateAssessmentReportIntegrityResponse_signatureAlgorithm :: Lens' ValidateAssessmentReportIntegrityResponse (Maybe Text)

The signature algorithm used to code sign the assessment report file.

validateAssessmentReportIntegrityResponse_signatureKeyId :: Lens' ValidateAssessmentReportIntegrityResponse (Maybe Text)

The unique identifier for the validation signature key.

validateAssessmentReportIntegrityResponse_httpStatus :: Lens' ValidateAssessmentReportIntegrityResponse Int

The response's http status code.