Amazonka.AutoScaling.BatchDeleteScheduledAction

Description

Deletes one or more scheduled actions for the specified Auto Scaling group.

Creating a Request

data BatchDeleteScheduledAction

See: newBatchDeleteScheduledAction smart constructor.

Constructors

BatchDeleteScheduledAction'

Fields

autoScalingGroupName :: Text

The name of the Auto Scaling group.

scheduledActionNames :: [Text]

The names of the scheduled actions to delete. The maximum number allowed is 50.

newBatchDeleteScheduledAction

Arguments

:: Text
$sel:autoScalingGroupName:BatchDeleteScheduledAction'

-> BatchDeleteScheduledAction

Create a value of BatchDeleteScheduledAction with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:autoScalingGroupName:BatchDeleteScheduledAction', batchDeleteScheduledAction_autoScalingGroupName - The name of the Auto Scaling group.

$sel:scheduledActionNames:BatchDeleteScheduledAction', batchDeleteScheduledAction_scheduledActionNames - The names of the scheduled actions to delete. The maximum number allowed is 50.

Request Lenses

batchDeleteScheduledAction_autoScalingGroupName :: Lens' BatchDeleteScheduledAction Text

The name of the Auto Scaling group.

batchDeleteScheduledAction_scheduledActionNames :: Lens' BatchDeleteScheduledAction [Text]

The names of the scheduled actions to delete. The maximum number allowed is 50.

Destructuring the Response

data BatchDeleteScheduledActionResponse

See: newBatchDeleteScheduledActionResponse smart constructor.

Constructors

BatchDeleteScheduledActionResponse'

Fields

failedScheduledActions :: Maybe [FailedScheduledUpdateGroupActionRequest]

The names of the scheduled actions that could not be deleted, including an error message.

httpStatus :: Int

The response's http status code.

newBatchDeleteScheduledActionResponse

Arguments

:: Int
$sel:httpStatus:BatchDeleteScheduledActionResponse'

-> BatchDeleteScheduledActionResponse

Create a value of BatchDeleteScheduledActionResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:failedScheduledActions:BatchDeleteScheduledActionResponse', batchDeleteScheduledActionResponse_failedScheduledActions - The names of the scheduled actions that could not be deleted, including an error message.

$sel:httpStatus:BatchDeleteScheduledActionResponse', batchDeleteScheduledActionResponse_httpStatus - The response's http status code.

Response Lenses

batchDeleteScheduledActionResponse_failedScheduledActions :: Lens' BatchDeleteScheduledActionResponse (Maybe [FailedScheduledUpdateGroupActionRequest])

The names of the scheduled actions that could not be deleted, including an error message.

batchDeleteScheduledActionResponse_httpStatus :: Lens' BatchDeleteScheduledActionResponse Int

The response's http status code.