Amazonka.AutoScaling.DescribeAutoScalingGroups

Description

Gets information about the Auto Scaling groups in the account and Region.

If you specify Auto Scaling group names, the output includes information for only the specified Auto Scaling groups. If you specify filters, the output includes information for only those Auto Scaling groups that meet the filter criteria. If you do not specify group names or filters, the output includes information for all Auto Scaling groups.

This operation also returns information about instances in Auto Scaling groups. To retrieve information about the instances in a warm pool, you must call the DescribeWarmPool API.

This operation returns paginated results.

Creating a Request

data DescribeAutoScalingGroups

See: newDescribeAutoScalingGroups smart constructor.

Constructors

DescribeAutoScalingGroups'

Fields

filters :: Maybe [Filter]

One or more filters to limit the results based on specific tags.

autoScalingGroupNames :: Maybe [Text]

The names of the Auto Scaling groups. By default, you can only specify up to 50 names. You can optionally increase this limit using the MaxRecords parameter.

If you omit this parameter, all Auto Scaling groups are described.

nextToken :: Maybe Text

The token for the next set of items to return. (You received this token from a previous call.)

maxRecords :: Maybe Int

The maximum number of items to return with this call. The default value is 50 and the maximum value is 100.

newDescribeAutoScalingGroups :: DescribeAutoScalingGroups

Create a value of DescribeAutoScalingGroups with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:filters:DescribeAutoScalingGroups', describeAutoScalingGroups_filters - One or more filters to limit the results based on specific tags.

$sel:autoScalingGroupNames:DescribeAutoScalingGroups', describeAutoScalingGroups_autoScalingGroupNames - The names of the Auto Scaling groups. By default, you can only specify up to 50 names. You can optionally increase this limit using the MaxRecords parameter.

If you omit this parameter, all Auto Scaling groups are described.

$sel:nextToken:DescribeAutoScalingGroups', describeAutoScalingGroups_nextToken - The token for the next set of items to return. (You received this token from a previous call.)

$sel:maxRecords:DescribeAutoScalingGroups', describeAutoScalingGroups_maxRecords - The maximum number of items to return with this call. The default value is 50 and the maximum value is 100.

Request Lenses

describeAutoScalingGroups_filters :: Lens' DescribeAutoScalingGroups (Maybe [Filter])

One or more filters to limit the results based on specific tags.

describeAutoScalingGroups_autoScalingGroupNames :: Lens' DescribeAutoScalingGroups (Maybe [Text])

The names of the Auto Scaling groups. By default, you can only specify up to 50 names. You can optionally increase this limit using the MaxRecords parameter.

If you omit this parameter, all Auto Scaling groups are described.

describeAutoScalingGroups_nextToken :: Lens' DescribeAutoScalingGroups (Maybe Text)

The token for the next set of items to return. (You received this token from a previous call.)

describeAutoScalingGroups_maxRecords :: Lens' DescribeAutoScalingGroups (Maybe Int)

The maximum number of items to return with this call. The default value is 50 and the maximum value is 100.

Destructuring the Response

data DescribeAutoScalingGroupsResponse

See: newDescribeAutoScalingGroupsResponse smart constructor.

Constructors

DescribeAutoScalingGroupsResponse'

Fields

nextToken :: Maybe Text

A string that indicates that the response contains more items than can be returned in a single response. To receive additional items, specify this string for the NextToken value when requesting the next set of items. This value is null when there are no more items to return.

httpStatus :: Int

The response's http status code.

autoScalingGroups :: [AutoScalingGroup]

The groups.

newDescribeAutoScalingGroupsResponse

Arguments

:: Int
$sel:httpStatus:DescribeAutoScalingGroupsResponse'

-> DescribeAutoScalingGroupsResponse

Create a value of DescribeAutoScalingGroupsResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:nextToken:DescribeAutoScalingGroups', describeAutoScalingGroupsResponse_nextToken - A string that indicates that the response contains more items than can be returned in a single response. To receive additional items, specify this string for the NextToken value when requesting the next set of items. This value is null when there are no more items to return.

$sel:httpStatus:DescribeAutoScalingGroupsResponse', describeAutoScalingGroupsResponse_httpStatus - The response's http status code.

$sel:autoScalingGroups:DescribeAutoScalingGroupsResponse', describeAutoScalingGroupsResponse_autoScalingGroups - The groups.

Response Lenses

describeAutoScalingGroupsResponse_nextToken :: Lens' DescribeAutoScalingGroupsResponse (Maybe Text)

A string that indicates that the response contains more items than can be returned in a single response. To receive additional items, specify this string for the NextToken value when requesting the next set of items. This value is null when there are no more items to return.

describeAutoScalingGroupsResponse_httpStatus :: Lens' DescribeAutoScalingGroupsResponse Int

The response's http status code.

describeAutoScalingGroupsResponse_autoScalingGroups :: Lens' DescribeAutoScalingGroupsResponse [AutoScalingGroup]

The groups.