Amazonka.AutoScaling.DescribeAutoScalingNotificationTypes

Description

Describes the notification types that are supported by Amazon EC2 Auto Scaling.

Creating a Request

data DescribeAutoScalingNotificationTypes

See: newDescribeAutoScalingNotificationTypes smart constructor.

Constructors

DescribeAutoScalingNotificationTypes'

newDescribeAutoScalingNotificationTypes :: DescribeAutoScalingNotificationTypes

Create a value of DescribeAutoScalingNotificationTypes with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

Destructuring the Response

data DescribeAutoScalingNotificationTypesResponse

See: newDescribeAutoScalingNotificationTypesResponse smart constructor.

Constructors

DescribeAutoScalingNotificationTypesResponse'

Fields

autoScalingNotificationTypes :: Maybe [Text]

The notification types.

httpStatus :: Int

The response's http status code.

newDescribeAutoScalingNotificationTypesResponse

Arguments

:: Int
$sel:httpStatus:DescribeAutoScalingNotificationTypesResponse'

-> DescribeAutoScalingNotificationTypesResponse

Create a value of DescribeAutoScalingNotificationTypesResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:autoScalingNotificationTypes:DescribeAutoScalingNotificationTypesResponse', describeAutoScalingNotificationTypesResponse_autoScalingNotificationTypes - The notification types.

$sel:httpStatus:DescribeAutoScalingNotificationTypesResponse', describeAutoScalingNotificationTypesResponse_httpStatus - The response's http status code.

Response Lenses

describeAutoScalingNotificationTypesResponse_autoScalingNotificationTypes :: Lens' DescribeAutoScalingNotificationTypesResponse (Maybe [Text])

The notification types.

describeAutoScalingNotificationTypesResponse_httpStatus :: Lens' DescribeAutoScalingNotificationTypesResponse Int

The response's http status code.