Amazonka.AutoScaling.DescribeInstanceRefreshes

Description

Gets information about the instance refreshes for the specified Auto Scaling group.

This operation is part of the instance refresh feature in Amazon EC2 Auto Scaling, which helps you update instances in your Auto Scaling group after you make configuration changes.

To help you determine the status of an instance refresh, this operation returns information about the instance refreshes you previously initiated, including their status, end time, the percentage of the instance refresh that is complete, and the number of instances remaining to update before the instance refresh is complete.

The following are the possible statuses:

Pending - The request was created, but the operation has not started.

InProgress - The operation is in progress.

Successful - The operation completed successfully.

Failed - The operation failed to complete. You can troubleshoot using the status reason and the scaling activities.

Cancelling - An ongoing operation is being cancelled. Cancellation does not roll back any replacements that have already been completed, but it prevents new replacements from being started.

Cancelled - The operation is cancelled.

Creating a Request

data DescribeInstanceRefreshes

See: newDescribeInstanceRefreshes smart constructor.

Constructors

DescribeInstanceRefreshes'

Fields

nextToken :: Maybe Text

The token for the next set of items to return. (You received this token from a previous call.)

maxRecords :: Maybe Int

The maximum number of items to return with this call. The default value is 50 and the maximum value is 100.

instanceRefreshIds :: Maybe [Text]

One or more instance refresh IDs.

autoScalingGroupName :: Text

The name of the Auto Scaling group.

newDescribeInstanceRefreshes

Arguments

:: Text
$sel:autoScalingGroupName:DescribeInstanceRefreshes'

-> DescribeInstanceRefreshes

Create a value of DescribeInstanceRefreshes with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:nextToken:DescribeInstanceRefreshes', describeInstanceRefreshes_nextToken - The token for the next set of items to return. (You received this token from a previous call.)

$sel:maxRecords:DescribeInstanceRefreshes', describeInstanceRefreshes_maxRecords - The maximum number of items to return with this call. The default value is 50 and the maximum value is 100.

$sel:instanceRefreshIds:DescribeInstanceRefreshes', describeInstanceRefreshes_instanceRefreshIds - One or more instance refresh IDs.

$sel:autoScalingGroupName:DescribeInstanceRefreshes', describeInstanceRefreshes_autoScalingGroupName - The name of the Auto Scaling group.

Request Lenses

describeInstanceRefreshes_nextToken :: Lens' DescribeInstanceRefreshes (Maybe Text)

The token for the next set of items to return. (You received this token from a previous call.)

describeInstanceRefreshes_maxRecords :: Lens' DescribeInstanceRefreshes (Maybe Int)

The maximum number of items to return with this call. The default value is 50 and the maximum value is 100.

describeInstanceRefreshes_instanceRefreshIds :: Lens' DescribeInstanceRefreshes (Maybe [Text])

One or more instance refresh IDs.

describeInstanceRefreshes_autoScalingGroupName :: Lens' DescribeInstanceRefreshes Text

The name of the Auto Scaling group.

Destructuring the Response

data DescribeInstanceRefreshesResponse

See: newDescribeInstanceRefreshesResponse smart constructor.

Constructors

DescribeInstanceRefreshesResponse'

Fields

nextToken :: Maybe Text

A string that indicates that the response contains more items than can be returned in a single response. To receive additional items, specify this string for the NextToken value when requesting the next set of items. This value is null when there are no more items to return.

instanceRefreshes :: Maybe [InstanceRefresh]

The instance refreshes for the specified group.

httpStatus :: Int

The response's http status code.

newDescribeInstanceRefreshesResponse

Arguments

:: Int
$sel:httpStatus:DescribeInstanceRefreshesResponse'

-> DescribeInstanceRefreshesResponse

Create a value of DescribeInstanceRefreshesResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:nextToken:DescribeInstanceRefreshes', describeInstanceRefreshesResponse_nextToken - A string that indicates that the response contains more items than can be returned in a single response. To receive additional items, specify this string for the NextToken value when requesting the next set of items. This value is null when there are no more items to return.

$sel:instanceRefreshes:DescribeInstanceRefreshesResponse', describeInstanceRefreshesResponse_instanceRefreshes - The instance refreshes for the specified group.

$sel:httpStatus:DescribeInstanceRefreshesResponse', describeInstanceRefreshesResponse_httpStatus - The response's http status code.

Response Lenses

describeInstanceRefreshesResponse_nextToken :: Lens' DescribeInstanceRefreshesResponse (Maybe Text)

A string that indicates that the response contains more items than can be returned in a single response. To receive additional items, specify this string for the NextToken value when requesting the next set of items. This value is null when there are no more items to return.

describeInstanceRefreshesResponse_instanceRefreshes :: Lens' DescribeInstanceRefreshesResponse (Maybe [InstanceRefresh])

The instance refreshes for the specified group.

describeInstanceRefreshesResponse_httpStatus :: Lens' DescribeInstanceRefreshesResponse Int

The response's http status code.