Amazonka.AutoScaling.DescribeLaunchConfigurations

Description

Gets information about the launch configurations in the account and Region.

This operation returns paginated results.

Creating a Request

data DescribeLaunchConfigurations

See: newDescribeLaunchConfigurations smart constructor.

Constructors

DescribeLaunchConfigurations'

Fields

launchConfigurationNames :: Maybe [Text]

The launch configuration names. If you omit this parameter, all launch configurations are described.

Array Members: Maximum number of 50 items.

nextToken :: Maybe Text

The token for the next set of items to return. (You received this token from a previous call.)

maxRecords :: Maybe Int

The maximum number of items to return with this call. The default value is 50 and the maximum value is 100.

newDescribeLaunchConfigurations :: DescribeLaunchConfigurations

Create a value of DescribeLaunchConfigurations with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:launchConfigurationNames:DescribeLaunchConfigurations', describeLaunchConfigurations_launchConfigurationNames - The launch configuration names. If you omit this parameter, all launch configurations are described.

Array Members: Maximum number of 50 items.

$sel:nextToken:DescribeLaunchConfigurations', describeLaunchConfigurations_nextToken - The token for the next set of items to return. (You received this token from a previous call.)

$sel:maxRecords:DescribeLaunchConfigurations', describeLaunchConfigurations_maxRecords - The maximum number of items to return with this call. The default value is 50 and the maximum value is 100.

Request Lenses

describeLaunchConfigurations_launchConfigurationNames :: Lens' DescribeLaunchConfigurations (Maybe [Text])

The launch configuration names. If you omit this parameter, all launch configurations are described.

Array Members: Maximum number of 50 items.

describeLaunchConfigurations_nextToken :: Lens' DescribeLaunchConfigurations (Maybe Text)

The token for the next set of items to return. (You received this token from a previous call.)

describeLaunchConfigurations_maxRecords :: Lens' DescribeLaunchConfigurations (Maybe Int)

The maximum number of items to return with this call. The default value is 50 and the maximum value is 100.

Destructuring the Response

data DescribeLaunchConfigurationsResponse

See: newDescribeLaunchConfigurationsResponse smart constructor.

Constructors

DescribeLaunchConfigurationsResponse'

Fields

nextToken :: Maybe Text

A string that indicates that the response contains more items than can be returned in a single response. To receive additional items, specify this string for the NextToken value when requesting the next set of items. This value is null when there are no more items to return.

httpStatus :: Int

The response's http status code.

launchConfigurations :: [LaunchConfiguration]

The launch configurations.

newDescribeLaunchConfigurationsResponse

Arguments

:: Int
$sel:httpStatus:DescribeLaunchConfigurationsResponse'

-> DescribeLaunchConfigurationsResponse

Create a value of DescribeLaunchConfigurationsResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:nextToken:DescribeLaunchConfigurations', describeLaunchConfigurationsResponse_nextToken - A string that indicates that the response contains more items than can be returned in a single response. To receive additional items, specify this string for the NextToken value when requesting the next set of items. This value is null when there are no more items to return.

$sel:httpStatus:DescribeLaunchConfigurationsResponse', describeLaunchConfigurationsResponse_httpStatus - The response's http status code.

$sel:launchConfigurations:DescribeLaunchConfigurationsResponse', describeLaunchConfigurationsResponse_launchConfigurations - The launch configurations.

Response Lenses

describeLaunchConfigurationsResponse_nextToken :: Lens' DescribeLaunchConfigurationsResponse (Maybe Text)

A string that indicates that the response contains more items than can be returned in a single response. To receive additional items, specify this string for the NextToken value when requesting the next set of items. This value is null when there are no more items to return.

describeLaunchConfigurationsResponse_httpStatus :: Lens' DescribeLaunchConfigurationsResponse Int

The response's http status code.

describeLaunchConfigurationsResponse_launchConfigurations :: Lens' DescribeLaunchConfigurationsResponse [LaunchConfiguration]

The launch configurations.