Amazonka.AutoScaling.DescribeLoadBalancerTargetGroups

Description

Gets information about the load balancer target groups for the specified Auto Scaling group.

To determine the availability of registered instances, use the State element in the response. When you attach a target group to an Auto Scaling group, the initial State value is Adding. The state transitions to Added after all Auto Scaling instances are registered with the target group. If Elastic Load Balancing health checks are enabled for the Auto Scaling group, the state transitions to InService after at least one Auto Scaling instance passes the health check. When the target group is in the InService state, Amazon EC2 Auto Scaling can terminate and replace any instances that are reported as unhealthy. If no registered instances pass the health checks, the target group doesn't enter the InService state.

Target groups also have an InService state if you attach them in the CreateAutoScalingGroup API call. If your target group state is InService, but it is not working properly, check the scaling activities by calling DescribeScalingActivities and take any corrective actions necessary.

For help with failed health checks, see Troubleshooting Amazon EC2 Auto Scaling: Health checks in the Amazon EC2 Auto Scaling User Guide. For more information, see Elastic Load Balancing and Amazon EC2 Auto Scaling in the Amazon EC2 Auto Scaling User Guide.

This operation returns paginated results.

Creating a Request

data DescribeLoadBalancerTargetGroups

See: newDescribeLoadBalancerTargetGroups smart constructor.

Constructors

DescribeLoadBalancerTargetGroups'

Fields

nextToken :: Maybe Text

The token for the next set of items to return. (You received this token from a previous call.)

maxRecords :: Maybe Int

The maximum number of items to return with this call. The default value is 100 and the maximum value is 100.

autoScalingGroupName :: Text

The name of the Auto Scaling group.

newDescribeLoadBalancerTargetGroups

Arguments

:: Text
$sel:autoScalingGroupName:DescribeLoadBalancerTargetGroups'

-> DescribeLoadBalancerTargetGroups

Create a value of DescribeLoadBalancerTargetGroups with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:nextToken:DescribeLoadBalancerTargetGroups', describeLoadBalancerTargetGroups_nextToken - The token for the next set of items to return. (You received this token from a previous call.)

$sel:maxRecords:DescribeLoadBalancerTargetGroups', describeLoadBalancerTargetGroups_maxRecords - The maximum number of items to return with this call. The default value is 100 and the maximum value is 100.

$sel:autoScalingGroupName:DescribeLoadBalancerTargetGroups', describeLoadBalancerTargetGroups_autoScalingGroupName - The name of the Auto Scaling group.

Request Lenses

describeLoadBalancerTargetGroups_nextToken :: Lens' DescribeLoadBalancerTargetGroups (Maybe Text)

The token for the next set of items to return. (You received this token from a previous call.)

describeLoadBalancerTargetGroups_maxRecords :: Lens' DescribeLoadBalancerTargetGroups (Maybe Int)

The maximum number of items to return with this call. The default value is 100 and the maximum value is 100.

describeLoadBalancerTargetGroups_autoScalingGroupName :: Lens' DescribeLoadBalancerTargetGroups Text

The name of the Auto Scaling group.

Destructuring the Response

data DescribeLoadBalancerTargetGroupsResponse

See: newDescribeLoadBalancerTargetGroupsResponse smart constructor.

Constructors

DescribeLoadBalancerTargetGroupsResponse'

Fields

loadBalancerTargetGroups :: Maybe [LoadBalancerTargetGroupState]

Information about the target groups.

nextToken :: Maybe Text

A string that indicates that the response contains more items than can be returned in a single response. To receive additional items, specify this string for the NextToken value when requesting the next set of items. This value is null when there are no more items to return.

httpStatus :: Int

The response's http status code.

newDescribeLoadBalancerTargetGroupsResponse

Arguments

:: Int
$sel:httpStatus:DescribeLoadBalancerTargetGroupsResponse'

-> DescribeLoadBalancerTargetGroupsResponse

Create a value of DescribeLoadBalancerTargetGroupsResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:loadBalancerTargetGroups:DescribeLoadBalancerTargetGroupsResponse', describeLoadBalancerTargetGroupsResponse_loadBalancerTargetGroups - Information about the target groups.

$sel:nextToken:DescribeLoadBalancerTargetGroups', describeLoadBalancerTargetGroupsResponse_nextToken - A string that indicates that the response contains more items than can be returned in a single response. To receive additional items, specify this string for the NextToken value when requesting the next set of items. This value is null when there are no more items to return.

$sel:httpStatus:DescribeLoadBalancerTargetGroupsResponse', describeLoadBalancerTargetGroupsResponse_httpStatus - The response's http status code.

Response Lenses

describeLoadBalancerTargetGroupsResponse_loadBalancerTargetGroups :: Lens' DescribeLoadBalancerTargetGroupsResponse (Maybe [LoadBalancerTargetGroupState])

Information about the target groups.

describeLoadBalancerTargetGroupsResponse_nextToken :: Lens' DescribeLoadBalancerTargetGroupsResponse (Maybe Text)

A string that indicates that the response contains more items than can be returned in a single response. To receive additional items, specify this string for the NextToken value when requesting the next set of items. This value is null when there are no more items to return.

describeLoadBalancerTargetGroupsResponse_httpStatus :: Lens' DescribeLoadBalancerTargetGroupsResponse Int

The response's http status code.