Amazonka.AutoScaling.EnableMetricsCollection

Description

Enables group metrics for the specified Auto Scaling group. For more information, see Monitoring CloudWatch metrics for your Auto Scaling groups and instances in the Amazon EC2 Auto Scaling User Guide.

Creating a Request

data EnableMetricsCollection

See: newEnableMetricsCollection smart constructor.

Constructors

EnableMetricsCollection'

Fields

metrics :: Maybe [Text]

Specifies which group-level metrics to start collecting. You can specify one or more of the following metrics:

GroupMinSize

GroupMaxSize

GroupDesiredCapacity

GroupInServiceInstances

GroupPendingInstances

GroupStandbyInstances

GroupTerminatingInstances

GroupTotalInstances

The instance weighting feature supports the following additional metrics:

GroupInServiceCapacity

GroupPendingCapacity

GroupStandbyCapacity

GroupTerminatingCapacity

GroupTotalCapacity

The warm pools feature supports the following additional metrics:

WarmPoolDesiredCapacity

WarmPoolWarmedCapacity

WarmPoolPendingCapacity

WarmPoolTerminatingCapacity

WarmPoolTotalCapacity

GroupAndWarmPoolDesiredCapacity

GroupAndWarmPoolTotalCapacity

If you omit this parameter, all metrics are enabled.

autoScalingGroupName :: Text

The name of the Auto Scaling group.

granularity :: Text

The granularity to associate with the metrics to collect. The only valid value is 1Minute.

newEnableMetricsCollection

Arguments

:: Text
$sel:autoScalingGroupName:EnableMetricsCollection'

-> Text
$sel:granularity:EnableMetricsCollection'

-> EnableMetricsCollection

Create a value of EnableMetricsCollection with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:metrics:EnableMetricsCollection', enableMetricsCollection_metrics - Specifies which group-level metrics to start collecting. You can specify one or more of the following metrics:

GroupMinSize

GroupMaxSize

GroupDesiredCapacity

GroupInServiceInstances

GroupPendingInstances

GroupStandbyInstances

GroupTerminatingInstances

GroupTotalInstances

The instance weighting feature supports the following additional metrics:

GroupInServiceCapacity

GroupPendingCapacity

GroupStandbyCapacity

GroupTerminatingCapacity

GroupTotalCapacity

The warm pools feature supports the following additional metrics:

WarmPoolDesiredCapacity

WarmPoolWarmedCapacity

WarmPoolPendingCapacity

WarmPoolTerminatingCapacity

WarmPoolTotalCapacity

GroupAndWarmPoolDesiredCapacity

GroupAndWarmPoolTotalCapacity

If you omit this parameter, all metrics are enabled.

$sel:autoScalingGroupName:EnableMetricsCollection', enableMetricsCollection_autoScalingGroupName - The name of the Auto Scaling group.

$sel:granularity:EnableMetricsCollection', enableMetricsCollection_granularity - The granularity to associate with the metrics to collect. The only valid value is 1Minute.

Request Lenses

enableMetricsCollection_metrics :: Lens' EnableMetricsCollection (Maybe [Text])

Specifies which group-level metrics to start collecting. You can specify one or more of the following metrics:

GroupMinSize

GroupMaxSize

GroupDesiredCapacity

GroupInServiceInstances

GroupPendingInstances

GroupStandbyInstances

GroupTerminatingInstances

GroupTotalInstances

The instance weighting feature supports the following additional metrics:

GroupInServiceCapacity

GroupPendingCapacity

GroupStandbyCapacity

GroupTerminatingCapacity

GroupTotalCapacity

The warm pools feature supports the following additional metrics:

WarmPoolDesiredCapacity

WarmPoolWarmedCapacity

WarmPoolPendingCapacity

WarmPoolTerminatingCapacity

WarmPoolTotalCapacity

GroupAndWarmPoolDesiredCapacity

GroupAndWarmPoolTotalCapacity

If you omit this parameter, all metrics are enabled.

enableMetricsCollection_autoScalingGroupName :: Lens' EnableMetricsCollection Text

The name of the Auto Scaling group.

enableMetricsCollection_granularity :: Lens' EnableMetricsCollection Text

The granularity to associate with the metrics to collect. The only valid value is 1Minute.

Destructuring the Response

data EnableMetricsCollectionResponse

See: newEnableMetricsCollectionResponse smart constructor.

Constructors

EnableMetricsCollectionResponse'

newEnableMetricsCollectionResponse :: EnableMetricsCollectionResponse

Create a value of EnableMetricsCollectionResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.