Amazonka.AutoScaling.PutLifecycleHook

Description

Creates or updates a lifecycle hook for the specified Auto Scaling group.

A lifecycle hook tells Amazon EC2 Auto Scaling to perform an action on an instance when the instance launches (before it is put into service) or as the instance terminates (before it is fully terminated).

This step is a part of the procedure for adding a lifecycle hook to an Auto Scaling group:

(Optional) Create a Lambda function and a rule that allows CloudWatch Events to invoke your Lambda function when Amazon EC2 Auto Scaling launches or terminates instances.

(Optional) Create a notification target and an IAM role. The target can be either an Amazon SQS queue or an Amazon SNS topic. The role allows Amazon EC2 Auto Scaling to publish lifecycle notifications to the target.

__Create the lifecycle hook. Specify whether the hook is used when the instances launch or terminate.__

If you need more time, record the lifecycle action heartbeat to keep the instance in a pending state using the RecordLifecycleActionHeartbeat API call.

If you finish before the timeout period ends, complete the lifecycle action using the CompleteLifecycleAction API call.

For more information, see Amazon EC2 Auto Scaling lifecycle hooks in the Amazon EC2 Auto Scaling User Guide.

If you exceed your maximum limit of lifecycle hooks, which by default is 50 per Auto Scaling group, the call fails.

You can view the lifecycle hooks for an Auto Scaling group using the DescribeLifecycleHooks API call. If you are no longer using a lifecycle hook, you can delete it by calling the DeleteLifecycleHook API.

Creating a Request

data PutLifecycleHook

See: newPutLifecycleHook smart constructor.

Constructors

PutLifecycleHook'

Fields

defaultResult :: Maybe Text

Defines the action the Auto Scaling group should take when the lifecycle hook timeout elapses or if an unexpected failure occurs. This parameter can be either CONTINUE or ABANDON. The default value is ABANDON.

heartbeatTimeout :: Maybe Int

The maximum time, in seconds, that can elapse before the lifecycle hook times out. The range is from 30 to 7200 seconds. The default value is 3600 seconds (1 hour).

If the lifecycle hook times out, Amazon EC2 Auto Scaling performs the action that you specified in the DefaultResult parameter. You can prevent the lifecycle hook from timing out by calling the RecordLifecycleActionHeartbeat API.

notificationMetadata :: Maybe Text

Additional information that you want to include any time Amazon EC2 Auto Scaling sends a message to the notification target.

notificationTargetARN :: Maybe Text

The ARN of the notification target that Amazon EC2 Auto Scaling uses to notify you when an instance is in the transition state for the lifecycle hook. This target can be either an SQS queue or an SNS topic.

If you specify an empty string, this overrides the current ARN.

This operation uses the JSON format when sending notifications to an Amazon SQS queue, and an email key-value pair format when sending notifications to an Amazon SNS topic.

When you specify a notification target, Amazon EC2 Auto Scaling sends it a test message. Test messages contain the following additional key-value pair: "Event": "autoscaling:TEST_NOTIFICATION".

lifecycleTransition :: Maybe Text

The instance state to which you want to attach the lifecycle hook. The valid values are:

autoscaling:EC2_INSTANCE_LAUNCHING

autoscaling:EC2_INSTANCE_TERMINATING

Required for new lifecycle hooks, but optional when updating existing hooks.

roleARN :: Maybe Text

The ARN of the IAM role that allows the Auto Scaling group to publish to the specified notification target, for example, an Amazon SNS topic or an Amazon SQS queue.

Required for new lifecycle hooks, but optional when updating existing hooks.

lifecycleHookName :: Text

The name of the lifecycle hook.

autoScalingGroupName :: Text

The name of the Auto Scaling group.

newPutLifecycleHook

Arguments

:: Text
$sel:lifecycleHookName:PutLifecycleHook'

-> Text
$sel:autoScalingGroupName:PutLifecycleHook'

-> PutLifecycleHook

Create a value of PutLifecycleHook with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:defaultResult:PutLifecycleHook', putLifecycleHook_defaultResult - Defines the action the Auto Scaling group should take when the lifecycle hook timeout elapses or if an unexpected failure occurs. This parameter can be either CONTINUE or ABANDON. The default value is ABANDON.

$sel:heartbeatTimeout:PutLifecycleHook', putLifecycleHook_heartbeatTimeout - The maximum time, in seconds, that can elapse before the lifecycle hook times out. The range is from 30 to 7200 seconds. The default value is 3600 seconds (1 hour).

If the lifecycle hook times out, Amazon EC2 Auto Scaling performs the action that you specified in the DefaultResult parameter. You can prevent the lifecycle hook from timing out by calling the RecordLifecycleActionHeartbeat API.

$sel:notificationMetadata:PutLifecycleHook', putLifecycleHook_notificationMetadata - Additional information that you want to include any time Amazon EC2 Auto Scaling sends a message to the notification target.

$sel:notificationTargetARN:PutLifecycleHook', putLifecycleHook_notificationTargetARN - The ARN of the notification target that Amazon EC2 Auto Scaling uses to notify you when an instance is in the transition state for the lifecycle hook. This target can be either an SQS queue or an SNS topic.

If you specify an empty string, this overrides the current ARN.

This operation uses the JSON format when sending notifications to an Amazon SQS queue, and an email key-value pair format when sending notifications to an Amazon SNS topic.

When you specify a notification target, Amazon EC2 Auto Scaling sends it a test message. Test messages contain the following additional key-value pair: "Event": "autoscaling:TEST_NOTIFICATION".

$sel:lifecycleTransition:PutLifecycleHook', putLifecycleHook_lifecycleTransition - The instance state to which you want to attach the lifecycle hook. The valid values are:

autoscaling:EC2_INSTANCE_LAUNCHING

autoscaling:EC2_INSTANCE_TERMINATING

Required for new lifecycle hooks, but optional when updating existing hooks.

$sel:roleARN:PutLifecycleHook', putLifecycleHook_roleARN - The ARN of the IAM role that allows the Auto Scaling group to publish to the specified notification target, for example, an Amazon SNS topic or an Amazon SQS queue.

Required for new lifecycle hooks, but optional when updating existing hooks.

$sel:lifecycleHookName:PutLifecycleHook', putLifecycleHook_lifecycleHookName - The name of the lifecycle hook.

$sel:autoScalingGroupName:PutLifecycleHook', putLifecycleHook_autoScalingGroupName - The name of the Auto Scaling group.

Request Lenses

putLifecycleHook_defaultResult :: Lens' PutLifecycleHook (Maybe Text)

Defines the action the Auto Scaling group should take when the lifecycle hook timeout elapses or if an unexpected failure occurs. This parameter can be either CONTINUE or ABANDON. The default value is ABANDON.

putLifecycleHook_heartbeatTimeout :: Lens' PutLifecycleHook (Maybe Int)

The maximum time, in seconds, that can elapse before the lifecycle hook times out. The range is from 30 to 7200 seconds. The default value is 3600 seconds (1 hour).

If the lifecycle hook times out, Amazon EC2 Auto Scaling performs the action that you specified in the DefaultResult parameter. You can prevent the lifecycle hook from timing out by calling the RecordLifecycleActionHeartbeat API.

putLifecycleHook_notificationMetadata :: Lens' PutLifecycleHook (Maybe Text)

Additional information that you want to include any time Amazon EC2 Auto Scaling sends a message to the notification target.

putLifecycleHook_notificationTargetARN :: Lens' PutLifecycleHook (Maybe Text)

The ARN of the notification target that Amazon EC2 Auto Scaling uses to notify you when an instance is in the transition state for the lifecycle hook. This target can be either an SQS queue or an SNS topic.

If you specify an empty string, this overrides the current ARN.

This operation uses the JSON format when sending notifications to an Amazon SQS queue, and an email key-value pair format when sending notifications to an Amazon SNS topic.

When you specify a notification target, Amazon EC2 Auto Scaling sends it a test message. Test messages contain the following additional key-value pair: "Event": "autoscaling:TEST_NOTIFICATION".

putLifecycleHook_lifecycleTransition :: Lens' PutLifecycleHook (Maybe Text)

The instance state to which you want to attach the lifecycle hook. The valid values are:

autoscaling:EC2_INSTANCE_LAUNCHING

autoscaling:EC2_INSTANCE_TERMINATING

Required for new lifecycle hooks, but optional when updating existing hooks.

putLifecycleHook_roleARN :: Lens' PutLifecycleHook (Maybe Text)

The ARN of the IAM role that allows the Auto Scaling group to publish to the specified notification target, for example, an Amazon SNS topic or an Amazon SQS queue.

Required for new lifecycle hooks, but optional when updating existing hooks.

putLifecycleHook_lifecycleHookName :: Lens' PutLifecycleHook Text

The name of the lifecycle hook.

putLifecycleHook_autoScalingGroupName :: Lens' PutLifecycleHook Text

The name of the Auto Scaling group.

Destructuring the Response

data PutLifecycleHookResponse

See: newPutLifecycleHookResponse smart constructor.

Constructors

PutLifecycleHookResponse'

Fields

httpStatus :: Int

The response's http status code.

newPutLifecycleHookResponse

Arguments

:: Int
$sel:httpStatus:PutLifecycleHookResponse'

-> PutLifecycleHookResponse

Create a value of PutLifecycleHookResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:httpStatus:PutLifecycleHookResponse', putLifecycleHookResponse_httpStatus - The response's http status code.

Response Lenses

putLifecycleHookResponse_httpStatus :: Lens' PutLifecycleHookResponse Int

The response's http status code.