Amazonka.AutoScaling.Types.PredictiveScalingPredefinedMetricPair

Description

Documentation

data PredictiveScalingPredefinedMetricPair

Represents a metric pair for a predictive scaling policy.

See: newPredictiveScalingPredefinedMetricPair smart constructor.

Constructors

PredictiveScalingPredefinedMetricPair'

Fields

resourceLabel :: Maybe Text

A label that uniquely identifies a specific Application Load Balancer target group from which to determine the total and average request count served by your Auto Scaling group. You can't specify a resource label unless the target group is attached to the Auto Scaling group.

You create the resource label by appending the final portion of the load balancer ARN and the final portion of the target group ARN into a single value, separated by a forward slash (/). The format of the resource label is:

app/my-alb/778d41231b141a0f/targetgroup/my-alb-target-group/943f017f100becff.

Where:

app/<load-balancer-name>/<load-balancer-id> is the final portion of the load balancer ARN

targetgroup/<target-group-name>/<target-group-id> is the final portion of the target group ARN.

To find the ARN for an Application Load Balancer, use the DescribeLoadBalancers API operation. To find the ARN for the target group, use the DescribeTargetGroups API operation.

predefinedMetricType :: PredefinedMetricPairType

Indicates which metrics to use. There are two different types of metrics for each metric type: one is a load metric and one is a scaling metric. For example, if the metric type is ASGCPUUtilization, the Auto Scaling group's total CPU metric is used as the load metric, and the average CPU metric is used for the scaling metric.

newPredictiveScalingPredefinedMetricPair

Arguments

:: PredefinedMetricPairType
$sel:predefinedMetricType:PredictiveScalingPredefinedMetricPair'

-> PredictiveScalingPredefinedMetricPair

Create a value of PredictiveScalingPredefinedMetricPair with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:resourceLabel:PredictiveScalingPredefinedMetricPair', predictiveScalingPredefinedMetricPair_resourceLabel - A label that uniquely identifies a specific Application Load Balancer target group from which to determine the total and average request count served by your Auto Scaling group. You can't specify a resource label unless the target group is attached to the Auto Scaling group.

You create the resource label by appending the final portion of the load balancer ARN and the final portion of the target group ARN into a single value, separated by a forward slash (/). The format of the resource label is:

app/my-alb/778d41231b141a0f/targetgroup/my-alb-target-group/943f017f100becff.

Where:

app/<load-balancer-name>/<load-balancer-id> is the final portion of the load balancer ARN

targetgroup/<target-group-name>/<target-group-id> is the final portion of the target group ARN.

To find the ARN for an Application Load Balancer, use the DescribeLoadBalancers API operation. To find the ARN for the target group, use the DescribeTargetGroups API operation.

$sel:predefinedMetricType:PredictiveScalingPredefinedMetricPair', predictiveScalingPredefinedMetricPair_predefinedMetricType - Indicates which metrics to use. There are two different types of metrics for each metric type: one is a load metric and one is a scaling metric. For example, if the metric type is ASGCPUUtilization, the Auto Scaling group's total CPU metric is used as the load metric, and the average CPU metric is used for the scaling metric.

predictiveScalingPredefinedMetricPair_resourceLabel :: Lens' PredictiveScalingPredefinedMetricPair (Maybe Text)

A label that uniquely identifies a specific Application Load Balancer target group from which to determine the total and average request count served by your Auto Scaling group. You can't specify a resource label unless the target group is attached to the Auto Scaling group.

You create the resource label by appending the final portion of the load balancer ARN and the final portion of the target group ARN into a single value, separated by a forward slash (/). The format of the resource label is:

app/my-alb/778d41231b141a0f/targetgroup/my-alb-target-group/943f017f100becff.

Where:

app/<load-balancer-name>/<load-balancer-id> is the final portion of the load balancer ARN

targetgroup/<target-group-name>/<target-group-id> is the final portion of the target group ARN.

To find the ARN for an Application Load Balancer, use the DescribeLoadBalancers API operation. To find the ARN for the target group, use the DescribeTargetGroups API operation.

predictiveScalingPredefinedMetricPair_predefinedMetricType :: Lens' PredictiveScalingPredefinedMetricPair PredefinedMetricPairType

Indicates which metrics to use. There are two different types of metrics for each metric type: one is a load metric and one is a scaling metric. For example, if the metric type is ASGCPUUtilization, the Auto Scaling group's total CPU metric is used as the load metric, and the average CPU metric is used for the scaling metric.