Amazonka.AutoScaling.Types.ScalingProcessQuery

Description

Documentation

data ScalingProcessQuery

See: newScalingProcessQuery smart constructor.

Constructors

ScalingProcessQuery'

Fields

scalingProcesses :: Maybe [Text]

One or more of the following processes:

Launch

Terminate

AddToLoadBalancer

AlarmNotification

AZRebalance

HealthCheck

InstanceRefresh

ReplaceUnhealthy

ScheduledActions

If you omit this parameter, all processes are specified.

autoScalingGroupName :: Text

The name of the Auto Scaling group.

newScalingProcessQuery

Arguments

:: Text
$sel:autoScalingGroupName:ScalingProcessQuery'

-> ScalingProcessQuery

Create a value of ScalingProcessQuery with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:scalingProcesses:ScalingProcessQuery', scalingProcessQuery_scalingProcesses - One or more of the following processes:

Launch

Terminate

AddToLoadBalancer

AlarmNotification

AZRebalance

HealthCheck

InstanceRefresh

ReplaceUnhealthy

ScheduledActions

If you omit this parameter, all processes are specified.

$sel:autoScalingGroupName:ScalingProcessQuery', scalingProcessQuery_autoScalingGroupName - The name of the Auto Scaling group.

scalingProcessQuery_scalingProcesses :: Lens' ScalingProcessQuery (Maybe [Text])

One or more of the following processes:

Launch

Terminate

AddToLoadBalancer

AlarmNotification

AZRebalance

HealthCheck

InstanceRefresh

ReplaceUnhealthy

ScheduledActions

If you omit this parameter, all processes are specified.

scalingProcessQuery_autoScalingGroupName :: Lens' ScalingProcessQuery Text

The name of the Auto Scaling group.