Amazonka.AutoScaling.Types.TargetTrackingConfiguration

Description

Documentation

data TargetTrackingConfiguration

Represents a target tracking scaling policy configuration to use with Amazon EC2 Auto Scaling.

See: newTargetTrackingConfiguration smart constructor.

Constructors

TargetTrackingConfiguration'

Fields

predefinedMetricSpecification :: Maybe PredefinedMetricSpecification

A predefined metric. You must specify either a predefined metric or a customized metric.

customizedMetricSpecification :: Maybe CustomizedMetricSpecification

A customized metric. You must specify either a predefined metric or a customized metric.

disableScaleIn :: Maybe Bool

Indicates whether scaling in by the target tracking scaling policy is disabled. If scaling in is disabled, the target tracking scaling policy doesn't remove instances from the Auto Scaling group. Otherwise, the target tracking scaling policy can remove instances from the Auto Scaling group. The default is false.

targetValue :: Double

The target value for the metric.

newTargetTrackingConfiguration

Arguments

:: Double
$sel:targetValue:TargetTrackingConfiguration'

-> TargetTrackingConfiguration

Create a value of TargetTrackingConfiguration with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:predefinedMetricSpecification:TargetTrackingConfiguration', targetTrackingConfiguration_predefinedMetricSpecification - A predefined metric. You must specify either a predefined metric or a customized metric.

$sel:customizedMetricSpecification:TargetTrackingConfiguration', targetTrackingConfiguration_customizedMetricSpecification - A customized metric. You must specify either a predefined metric or a customized metric.

$sel:disableScaleIn:TargetTrackingConfiguration', targetTrackingConfiguration_disableScaleIn - Indicates whether scaling in by the target tracking scaling policy is disabled. If scaling in is disabled, the target tracking scaling policy doesn't remove instances from the Auto Scaling group. Otherwise, the target tracking scaling policy can remove instances from the Auto Scaling group. The default is false.

$sel:targetValue:TargetTrackingConfiguration', targetTrackingConfiguration_targetValue - The target value for the metric.

targetTrackingConfiguration_predefinedMetricSpecification :: Lens' TargetTrackingConfiguration (Maybe PredefinedMetricSpecification)

A predefined metric. You must specify either a predefined metric or a customized metric.

targetTrackingConfiguration_customizedMetricSpecification :: Lens' TargetTrackingConfiguration (Maybe CustomizedMetricSpecification)

A customized metric. You must specify either a predefined metric or a customized metric.

targetTrackingConfiguration_disableScaleIn :: Lens' TargetTrackingConfiguration (Maybe Bool)

Indicates whether scaling in by the target tracking scaling policy is disabled. If scaling in is disabled, the target tracking scaling policy doesn't remove instances from the Auto Scaling group. Otherwise, the target tracking scaling policy can remove instances from the Auto Scaling group. The default is false.

targetTrackingConfiguration_targetValue :: Lens' TargetTrackingConfiguration Double

The target value for the metric.