Amazonka.AutoScalingPlans.Types.TargetTrackingConfiguration

Description

Documentation

data TargetTrackingConfiguration

Describes a target tracking configuration to use with AWS Auto Scaling. Used with ScalingInstruction and ScalingPolicy.

See: newTargetTrackingConfiguration smart constructor.

Constructors

TargetTrackingConfiguration'

Fields

estimatedInstanceWarmup :: Maybe Int

The estimated time, in seconds, until a newly launched instance can contribute to the CloudWatch metrics. This value is used only if the resource is an Auto Scaling group.

predefinedScalingMetricSpecification :: Maybe PredefinedScalingMetricSpecification

A predefined metric. You can specify either a predefined metric or a customized metric.

scaleInCooldown :: Maybe Int

The amount of time, in seconds, after a scale-in activity completes before another scale-in activity can start. This property is not used if the scalable resource is an Auto Scaling group.

With the scale-in cooldown period, the intention is to scale in conservatively to protect your application’s availability, so scale-in activities are blocked until the cooldown period has expired. However, if another alarm triggers a scale-out activity during the scale-in cooldown period, Auto Scaling scales out the target immediately. In this case, the scale-in cooldown period stops and doesn't complete.

disableScaleIn :: Maybe Bool

Indicates whether scale in by the target tracking scaling policy is disabled. If the value is true, scale in is disabled and the target tracking scaling policy doesn't remove capacity from the scalable resource. Otherwise, scale in is enabled and the target tracking scaling policy can remove capacity from the scalable resource.

The default value is false.

customizedScalingMetricSpecification :: Maybe CustomizedScalingMetricSpecification

A customized metric. You can specify either a predefined metric or a customized metric.

scaleOutCooldown :: Maybe Int

The amount of time, in seconds, to wait for a previous scale-out activity to take effect. This property is not used if the scalable resource is an Auto Scaling group.

With the scale-out cooldown period, the intention is to continuously (but not excessively) scale out. After Auto Scaling successfully scales out using a target tracking scaling policy, it starts to calculate the cooldown time. The scaling policy won't increase the desired capacity again unless either a larger scale out is triggered or the cooldown period ends.

targetValue :: Double

The target value for the metric. Although this property accepts numbers of type Double, it won't accept values that are either too small or too large. Values must be in the range of -2^360 to 2^360.

newTargetTrackingConfiguration

Arguments

:: Double
$sel:targetValue:TargetTrackingConfiguration'

-> TargetTrackingConfiguration

Create a value of TargetTrackingConfiguration with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:estimatedInstanceWarmup:TargetTrackingConfiguration', targetTrackingConfiguration_estimatedInstanceWarmup - The estimated time, in seconds, until a newly launched instance can contribute to the CloudWatch metrics. This value is used only if the resource is an Auto Scaling group.

$sel:predefinedScalingMetricSpecification:TargetTrackingConfiguration', targetTrackingConfiguration_predefinedScalingMetricSpecification - A predefined metric. You can specify either a predefined metric or a customized metric.

$sel:scaleInCooldown:TargetTrackingConfiguration', targetTrackingConfiguration_scaleInCooldown - The amount of time, in seconds, after a scale-in activity completes before another scale-in activity can start. This property is not used if the scalable resource is an Auto Scaling group.

With the scale-in cooldown period, the intention is to scale in conservatively to protect your application’s availability, so scale-in activities are blocked until the cooldown period has expired. However, if another alarm triggers a scale-out activity during the scale-in cooldown period, Auto Scaling scales out the target immediately. In this case, the scale-in cooldown period stops and doesn't complete.

$sel:disableScaleIn:TargetTrackingConfiguration', targetTrackingConfiguration_disableScaleIn - Indicates whether scale in by the target tracking scaling policy is disabled. If the value is true, scale in is disabled and the target tracking scaling policy doesn't remove capacity from the scalable resource. Otherwise, scale in is enabled and the target tracking scaling policy can remove capacity from the scalable resource.

The default value is false.

$sel:customizedScalingMetricSpecification:TargetTrackingConfiguration', targetTrackingConfiguration_customizedScalingMetricSpecification - A customized metric. You can specify either a predefined metric or a customized metric.

$sel:scaleOutCooldown:TargetTrackingConfiguration', targetTrackingConfiguration_scaleOutCooldown - The amount of time, in seconds, to wait for a previous scale-out activity to take effect. This property is not used if the scalable resource is an Auto Scaling group.

With the scale-out cooldown period, the intention is to continuously (but not excessively) scale out. After Auto Scaling successfully scales out using a target tracking scaling policy, it starts to calculate the cooldown time. The scaling policy won't increase the desired capacity again unless either a larger scale out is triggered or the cooldown period ends.

$sel:targetValue:TargetTrackingConfiguration', targetTrackingConfiguration_targetValue - The target value for the metric. Although this property accepts numbers of type Double, it won't accept values that are either too small or too large. Values must be in the range of -2^360 to 2^360.

targetTrackingConfiguration_estimatedInstanceWarmup :: Lens' TargetTrackingConfiguration (Maybe Int)

The estimated time, in seconds, until a newly launched instance can contribute to the CloudWatch metrics. This value is used only if the resource is an Auto Scaling group.

targetTrackingConfiguration_predefinedScalingMetricSpecification :: Lens' TargetTrackingConfiguration (Maybe PredefinedScalingMetricSpecification)

A predefined metric. You can specify either a predefined metric or a customized metric.

targetTrackingConfiguration_scaleInCooldown :: Lens' TargetTrackingConfiguration (Maybe Int)

The amount of time, in seconds, after a scale-in activity completes before another scale-in activity can start. This property is not used if the scalable resource is an Auto Scaling group.

With the scale-in cooldown period, the intention is to scale in conservatively to protect your application’s availability, so scale-in activities are blocked until the cooldown period has expired. However, if another alarm triggers a scale-out activity during the scale-in cooldown period, Auto Scaling scales out the target immediately. In this case, the scale-in cooldown period stops and doesn't complete.

targetTrackingConfiguration_disableScaleIn :: Lens' TargetTrackingConfiguration (Maybe Bool)

Indicates whether scale in by the target tracking scaling policy is disabled. If the value is true, scale in is disabled and the target tracking scaling policy doesn't remove capacity from the scalable resource. Otherwise, scale in is enabled and the target tracking scaling policy can remove capacity from the scalable resource.

The default value is false.

targetTrackingConfiguration_customizedScalingMetricSpecification :: Lens' TargetTrackingConfiguration (Maybe CustomizedScalingMetricSpecification)

A customized metric. You can specify either a predefined metric or a customized metric.

targetTrackingConfiguration_scaleOutCooldown :: Lens' TargetTrackingConfiguration (Maybe Int)

The amount of time, in seconds, to wait for a previous scale-out activity to take effect. This property is not used if the scalable resource is an Auto Scaling group.

With the scale-out cooldown period, the intention is to continuously (but not excessively) scale out. After Auto Scaling successfully scales out using a target tracking scaling policy, it starts to calculate the cooldown time. The scaling policy won't increase the desired capacity again unless either a larger scale out is triggered or the cooldown period ends.

targetTrackingConfiguration_targetValue :: Lens' TargetTrackingConfiguration Double

The target value for the metric. Although this property accepts numbers of type Double, it won't accept values that are either too small or too large. Values must be in the range of -2^360 to 2^360.