Amazonka.Backup.DescribeBackupJob

Description

Returns backup job details for the specified BackupJobId.

Creating a Request

data DescribeBackupJob

See: newDescribeBackupJob smart constructor.

Constructors

DescribeBackupJob'

Fields

backupJobId :: Text

Uniquely identifies a request to Backup to back up a resource.

newDescribeBackupJob

Arguments

:: Text
$sel:backupJobId:DescribeBackupJob'

-> DescribeBackupJob

Create a value of DescribeBackupJob with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:backupJobId:DescribeBackupJob', describeBackupJob_backupJobId - Uniquely identifies a request to Backup to back up a resource.

Request Lenses

describeBackupJob_backupJobId :: Lens' DescribeBackupJob Text

Uniquely identifies a request to Backup to back up a resource.

Destructuring the Response

data DescribeBackupJobResponse

See: newDescribeBackupJobResponse smart constructor.

Constructors

DescribeBackupJobResponse'

Fields

iamRoleArn :: Maybe Text

Specifies the IAM role ARN used to create the target recovery point; for example, arn:aws:iam::123456789012:role/S3Access.

state :: Maybe BackupJobState

The current state of a resource recovery point.

resourceType :: Maybe Text

The type of Amazon Web Services resource to be backed up; for example, an Amazon Elastic Block Store (Amazon EBS) volume or an Amazon Relational Database Service (Amazon RDS) database.

percentDone :: Maybe Text

Contains an estimated percentage that is complete of a job at the time the job status was queried.

startBy :: Maybe POSIX

Specifies the time in Unix format and Coordinated Universal Time (UTC) when a backup job must be started before it is canceled. The value is calculated by adding the start window to the scheduled time. So if the scheduled time were 6:00 PM and the start window is 2 hours, the StartBy time would be 8:00 PM on the date specified. The value of StartBy is accurate to milliseconds. For example, the value 1516925490.087 represents Friday, January 26, 2018 12:11:30.087 AM.

createdBy :: Maybe RecoveryPointCreator

Contains identifying information about the creation of a backup job, including the BackupPlanArn, BackupPlanId, BackupPlanVersion, and BackupRuleId of the backup plan that is used to create it.

expectedCompletionDate :: Maybe POSIX

The date and time that a job to back up resources is expected to be completed, in Unix format and Coordinated Universal Time (UTC). The value of ExpectedCompletionDate is accurate to milliseconds. For example, the value 1516925490.087 represents Friday, January 26, 2018 12:11:30.087 AM.

bytesTransferred :: Maybe Integer

The size in bytes transferred to a backup vault at the time that the job status was queried.

backupVaultArn :: Maybe Text

An Amazon Resource Name (ARN) that uniquely identifies a backup vault; for example, arn:aws:backup:us-east-1:123456789012:vault:aBackupVault.

accountId :: Maybe Text

Returns the account ID that owns the backup job.

backupJobId :: Maybe Text

Uniquely identifies a request to Backup to back up a resource.

resourceArn :: Maybe Text

An ARN that uniquely identifies a saved resource. The format of the ARN depends on the resource type.

statusMessage :: Maybe Text

A detailed message explaining the status of the job to back up a resource.

recoveryPointArn :: Maybe Text

An ARN that uniquely identifies a recovery point; for example, arn:aws:backup:us-east-1:123456789012:recovery-point:1EB3B5E7-9EB0-435A-A80B-108B488B0D45.

backupSizeInBytes :: Maybe Integer

The size, in bytes, of a backup.

creationDate :: Maybe POSIX

The date and time that a backup job is created, in Unix format and Coordinated Universal Time (UTC). The value of CreationDate is accurate to milliseconds. For example, the value 1516925490.087 represents Friday, January 26, 2018 12:11:30.087 AM.

completionDate :: Maybe POSIX

The date and time that a job to create a backup job is completed, in Unix format and Coordinated Universal Time (UTC). The value of CompletionDate is accurate to milliseconds. For example, the value 1516925490.087 represents Friday, January 26, 2018 12:11:30.087 AM.

backupVaultName :: Maybe Text

The name of a logical container where backups are stored. Backup vaults are identified by names that are unique to the account used to create them and the Amazon Web Services Region where they are created. They consist of lowercase letters, numbers, and hyphens.

backupType :: Maybe Text

Represents the actual backup type selected for a backup job. For example, if a successful Windows Volume Shadow Copy Service (VSS) backup was taken, BackupType returns "WindowsVSS". If BackupType is empty, then the backup type was a regular backup.

backupOptions :: Maybe (HashMap Text Text)

Represents the options specified as part of backup plan or on-demand backup job.

httpStatus :: Int

The response's http status code.

newDescribeBackupJobResponse

Arguments

:: Int
$sel:httpStatus:DescribeBackupJobResponse'

-> DescribeBackupJobResponse

Create a value of DescribeBackupJobResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:iamRoleArn:DescribeBackupJobResponse', describeBackupJobResponse_iamRoleArn - Specifies the IAM role ARN used to create the target recovery point; for example, arn:aws:iam::123456789012:role/S3Access.

$sel:state:DescribeBackupJobResponse', describeBackupJobResponse_state - The current state of a resource recovery point.

$sel:resourceType:DescribeBackupJobResponse', describeBackupJobResponse_resourceType - The type of Amazon Web Services resource to be backed up; for example, an Amazon Elastic Block Store (Amazon EBS) volume or an Amazon Relational Database Service (Amazon RDS) database.

$sel:percentDone:DescribeBackupJobResponse', describeBackupJobResponse_percentDone - Contains an estimated percentage that is complete of a job at the time the job status was queried.

$sel:startBy:DescribeBackupJobResponse', describeBackupJobResponse_startBy - Specifies the time in Unix format and Coordinated Universal Time (UTC) when a backup job must be started before it is canceled. The value is calculated by adding the start window to the scheduled time. So if the scheduled time were 6:00 PM and the start window is 2 hours, the StartBy time would be 8:00 PM on the date specified. The value of StartBy is accurate to milliseconds. For example, the value 1516925490.087 represents Friday, January 26, 2018 12:11:30.087 AM.

$sel:createdBy:DescribeBackupJobResponse', describeBackupJobResponse_createdBy - Contains identifying information about the creation of a backup job, including the BackupPlanArn, BackupPlanId, BackupPlanVersion, and BackupRuleId of the backup plan that is used to create it.

$sel:expectedCompletionDate:DescribeBackupJobResponse', describeBackupJobResponse_expectedCompletionDate - The date and time that a job to back up resources is expected to be completed, in Unix format and Coordinated Universal Time (UTC). The value of ExpectedCompletionDate is accurate to milliseconds. For example, the value 1516925490.087 represents Friday, January 26, 2018 12:11:30.087 AM.

$sel:bytesTransferred:DescribeBackupJobResponse', describeBackupJobResponse_bytesTransferred - The size in bytes transferred to a backup vault at the time that the job status was queried.

$sel:backupVaultArn:DescribeBackupJobResponse', describeBackupJobResponse_backupVaultArn - An Amazon Resource Name (ARN) that uniquely identifies a backup vault; for example, arn:aws:backup:us-east-1:123456789012:vault:aBackupVault.

$sel:accountId:DescribeBackupJobResponse', describeBackupJobResponse_accountId - Returns the account ID that owns the backup job.

$sel:backupJobId:DescribeBackupJob', describeBackupJobResponse_backupJobId - Uniquely identifies a request to Backup to back up a resource.

$sel:resourceArn:DescribeBackupJobResponse', describeBackupJobResponse_resourceArn - An ARN that uniquely identifies a saved resource. The format of the ARN depends on the resource type.

$sel:statusMessage:DescribeBackupJobResponse', describeBackupJobResponse_statusMessage - A detailed message explaining the status of the job to back up a resource.

$sel:recoveryPointArn:DescribeBackupJobResponse', describeBackupJobResponse_recoveryPointArn - An ARN that uniquely identifies a recovery point; for example, arn:aws:backup:us-east-1:123456789012:recovery-point:1EB3B5E7-9EB0-435A-A80B-108B488B0D45.

$sel:backupSizeInBytes:DescribeBackupJobResponse', describeBackupJobResponse_backupSizeInBytes - The size, in bytes, of a backup.

$sel:creationDate:DescribeBackupJobResponse', describeBackupJobResponse_creationDate - The date and time that a backup job is created, in Unix format and Coordinated Universal Time (UTC). The value of CreationDate is accurate to milliseconds. For example, the value 1516925490.087 represents Friday, January 26, 2018 12:11:30.087 AM.

$sel:completionDate:DescribeBackupJobResponse', describeBackupJobResponse_completionDate - The date and time that a job to create a backup job is completed, in Unix format and Coordinated Universal Time (UTC). The value of CompletionDate is accurate to milliseconds. For example, the value 1516925490.087 represents Friday, January 26, 2018 12:11:30.087 AM.

$sel:backupVaultName:DescribeBackupJobResponse', describeBackupJobResponse_backupVaultName - The name of a logical container where backups are stored. Backup vaults are identified by names that are unique to the account used to create them and the Amazon Web Services Region where they are created. They consist of lowercase letters, numbers, and hyphens.

$sel:backupType:DescribeBackupJobResponse', describeBackupJobResponse_backupType - Represents the actual backup type selected for a backup job. For example, if a successful Windows Volume Shadow Copy Service (VSS) backup was taken, BackupType returns "WindowsVSS". If BackupType is empty, then the backup type was a regular backup.

$sel:backupOptions:DescribeBackupJobResponse', describeBackupJobResponse_backupOptions - Represents the options specified as part of backup plan or on-demand backup job.

$sel:httpStatus:DescribeBackupJobResponse', describeBackupJobResponse_httpStatus - The response's http status code.

Response Lenses

describeBackupJobResponse_iamRoleArn :: Lens' DescribeBackupJobResponse (Maybe Text)

Specifies the IAM role ARN used to create the target recovery point; for example, arn:aws:iam::123456789012:role/S3Access.

describeBackupJobResponse_state :: Lens' DescribeBackupJobResponse (Maybe BackupJobState)

The current state of a resource recovery point.

describeBackupJobResponse_resourceType :: Lens' DescribeBackupJobResponse (Maybe Text)

The type of Amazon Web Services resource to be backed up; for example, an Amazon Elastic Block Store (Amazon EBS) volume or an Amazon Relational Database Service (Amazon RDS) database.

describeBackupJobResponse_percentDone :: Lens' DescribeBackupJobResponse (Maybe Text)

Contains an estimated percentage that is complete of a job at the time the job status was queried.

describeBackupJobResponse_startBy :: Lens' DescribeBackupJobResponse (Maybe UTCTime)

Specifies the time in Unix format and Coordinated Universal Time (UTC) when a backup job must be started before it is canceled. The value is calculated by adding the start window to the scheduled time. So if the scheduled time were 6:00 PM and the start window is 2 hours, the StartBy time would be 8:00 PM on the date specified. The value of StartBy is accurate to milliseconds. For example, the value 1516925490.087 represents Friday, January 26, 2018 12:11:30.087 AM.

describeBackupJobResponse_createdBy :: Lens' DescribeBackupJobResponse (Maybe RecoveryPointCreator)

Contains identifying information about the creation of a backup job, including the BackupPlanArn, BackupPlanId, BackupPlanVersion, and BackupRuleId of the backup plan that is used to create it.

describeBackupJobResponse_expectedCompletionDate :: Lens' DescribeBackupJobResponse (Maybe UTCTime)

The date and time that a job to back up resources is expected to be completed, in Unix format and Coordinated Universal Time (UTC). The value of ExpectedCompletionDate is accurate to milliseconds. For example, the value 1516925490.087 represents Friday, January 26, 2018 12:11:30.087 AM.

describeBackupJobResponse_bytesTransferred :: Lens' DescribeBackupJobResponse (Maybe Integer)

The size in bytes transferred to a backup vault at the time that the job status was queried.

describeBackupJobResponse_backupVaultArn :: Lens' DescribeBackupJobResponse (Maybe Text)

An Amazon Resource Name (ARN) that uniquely identifies a backup vault; for example, arn:aws:backup:us-east-1:123456789012:vault:aBackupVault.

describeBackupJobResponse_accountId :: Lens' DescribeBackupJobResponse (Maybe Text)

Returns the account ID that owns the backup job.

describeBackupJobResponse_backupJobId :: Lens' DescribeBackupJobResponse (Maybe Text)

Uniquely identifies a request to Backup to back up a resource.

describeBackupJobResponse_resourceArn :: Lens' DescribeBackupJobResponse (Maybe Text)

An ARN that uniquely identifies a saved resource. The format of the ARN depends on the resource type.

describeBackupJobResponse_statusMessage :: Lens' DescribeBackupJobResponse (Maybe Text)

A detailed message explaining the status of the job to back up a resource.

describeBackupJobResponse_recoveryPointArn :: Lens' DescribeBackupJobResponse (Maybe Text)

An ARN that uniquely identifies a recovery point; for example, arn:aws:backup:us-east-1:123456789012:recovery-point:1EB3B5E7-9EB0-435A-A80B-108B488B0D45.

describeBackupJobResponse_backupSizeInBytes :: Lens' DescribeBackupJobResponse (Maybe Integer)

The size, in bytes, of a backup.

describeBackupJobResponse_creationDate :: Lens' DescribeBackupJobResponse (Maybe UTCTime)

The date and time that a backup job is created, in Unix format and Coordinated Universal Time (UTC). The value of CreationDate is accurate to milliseconds. For example, the value 1516925490.087 represents Friday, January 26, 2018 12:11:30.087 AM.

describeBackupJobResponse_completionDate :: Lens' DescribeBackupJobResponse (Maybe UTCTime)

The date and time that a job to create a backup job is completed, in Unix format and Coordinated Universal Time (UTC). The value of CompletionDate is accurate to milliseconds. For example, the value 1516925490.087 represents Friday, January 26, 2018 12:11:30.087 AM.

describeBackupJobResponse_backupVaultName :: Lens' DescribeBackupJobResponse (Maybe Text)

The name of a logical container where backups are stored. Backup vaults are identified by names that are unique to the account used to create them and the Amazon Web Services Region where they are created. They consist of lowercase letters, numbers, and hyphens.

describeBackupJobResponse_backupType :: Lens' DescribeBackupJobResponse (Maybe Text)

Represents the actual backup type selected for a backup job. For example, if a successful Windows Volume Shadow Copy Service (VSS) backup was taken, BackupType returns "WindowsVSS". If BackupType is empty, then the backup type was a regular backup.

describeBackupJobResponse_backupOptions :: Lens' DescribeBackupJobResponse (Maybe (HashMap Text Text))

Represents the options specified as part of backup plan or on-demand backup job.

describeBackupJobResponse_httpStatus :: Lens' DescribeBackupJobResponse Int

The response's http status code.