Amazonka.Backup.GetBackupVaultAccessPolicy

Description

Returns the access policy document that is associated with the named backup vault.

Creating a Request

data GetBackupVaultAccessPolicy

See: newGetBackupVaultAccessPolicy smart constructor.

Constructors

GetBackupVaultAccessPolicy'

Fields

backupVaultName :: Text

The name of a logical container where backups are stored. Backup vaults are identified by names that are unique to the account used to create them and the Amazon Web Services Region where they are created. They consist of lowercase letters, numbers, and hyphens.

newGetBackupVaultAccessPolicy

Arguments

:: Text
$sel:backupVaultName:GetBackupVaultAccessPolicy'

-> GetBackupVaultAccessPolicy

Create a value of GetBackupVaultAccessPolicy with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:backupVaultName:GetBackupVaultAccessPolicy', getBackupVaultAccessPolicy_backupVaultName - The name of a logical container where backups are stored. Backup vaults are identified by names that are unique to the account used to create them and the Amazon Web Services Region where they are created. They consist of lowercase letters, numbers, and hyphens.

Request Lenses

getBackupVaultAccessPolicy_backupVaultName :: Lens' GetBackupVaultAccessPolicy Text

The name of a logical container where backups are stored. Backup vaults are identified by names that are unique to the account used to create them and the Amazon Web Services Region where they are created. They consist of lowercase letters, numbers, and hyphens.

Destructuring the Response

data GetBackupVaultAccessPolicyResponse

See: newGetBackupVaultAccessPolicyResponse smart constructor.

Constructors

GetBackupVaultAccessPolicyResponse'

Fields

backupVaultArn :: Maybe Text

An Amazon Resource Name (ARN) that uniquely identifies a backup vault; for example, arn:aws:backup:us-east-1:123456789012:vault:aBackupVault.

policy :: Maybe Text

The backup vault access policy document in JSON format.

backupVaultName :: Maybe Text

The name of a logical container where backups are stored. Backup vaults are identified by names that are unique to the account used to create them and the Region where they are created. They consist of lowercase letters, numbers, and hyphens.

httpStatus :: Int

The response's http status code.

newGetBackupVaultAccessPolicyResponse

Arguments

:: Int
$sel:httpStatus:GetBackupVaultAccessPolicyResponse'

-> GetBackupVaultAccessPolicyResponse

Create a value of GetBackupVaultAccessPolicyResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:backupVaultArn:GetBackupVaultAccessPolicyResponse', getBackupVaultAccessPolicyResponse_backupVaultArn - An Amazon Resource Name (ARN) that uniquely identifies a backup vault; for example, arn:aws:backup:us-east-1:123456789012:vault:aBackupVault.

$sel:policy:GetBackupVaultAccessPolicyResponse', getBackupVaultAccessPolicyResponse_policy - The backup vault access policy document in JSON format.

$sel:backupVaultName:GetBackupVaultAccessPolicy', getBackupVaultAccessPolicyResponse_backupVaultName - The name of a logical container where backups are stored. Backup vaults are identified by names that are unique to the account used to create them and the Region where they are created. They consist of lowercase letters, numbers, and hyphens.

$sel:httpStatus:GetBackupVaultAccessPolicyResponse', getBackupVaultAccessPolicyResponse_httpStatus - The response's http status code.

Response Lenses

getBackupVaultAccessPolicyResponse_backupVaultArn :: Lens' GetBackupVaultAccessPolicyResponse (Maybe Text)

An Amazon Resource Name (ARN) that uniquely identifies a backup vault; for example, arn:aws:backup:us-east-1:123456789012:vault:aBackupVault.

getBackupVaultAccessPolicyResponse_policy :: Lens' GetBackupVaultAccessPolicyResponse (Maybe Text)

The backup vault access policy document in JSON format.

getBackupVaultAccessPolicyResponse_backupVaultName :: Lens' GetBackupVaultAccessPolicyResponse (Maybe Text)

The name of a logical container where backups are stored. Backup vaults are identified by names that are unique to the account used to create them and the Region where they are created. They consist of lowercase letters, numbers, and hyphens.

getBackupVaultAccessPolicyResponse_httpStatus :: Lens' GetBackupVaultAccessPolicyResponse Int

The response's http status code.