Amazonka.Backup.GetBackupVaultNotifications

Description

Returns event notifications for the specified backup vault.

Creating a Request

data GetBackupVaultNotifications

See: newGetBackupVaultNotifications smart constructor.

Constructors

GetBackupVaultNotifications'

Fields

backupVaultName :: Text

The name of a logical container where backups are stored. Backup vaults are identified by names that are unique to the account used to create them and the Amazon Web Services Region where they are created. They consist of lowercase letters, numbers, and hyphens.

newGetBackupVaultNotifications

Arguments

:: Text
$sel:backupVaultName:GetBackupVaultNotifications'

-> GetBackupVaultNotifications

Create a value of GetBackupVaultNotifications with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:backupVaultName:GetBackupVaultNotifications', getBackupVaultNotifications_backupVaultName - The name of a logical container where backups are stored. Backup vaults are identified by names that are unique to the account used to create them and the Amazon Web Services Region where they are created. They consist of lowercase letters, numbers, and hyphens.

Request Lenses

getBackupVaultNotifications_backupVaultName :: Lens' GetBackupVaultNotifications Text

The name of a logical container where backups are stored. Backup vaults are identified by names that are unique to the account used to create them and the Amazon Web Services Region where they are created. They consist of lowercase letters, numbers, and hyphens.

Destructuring the Response

data GetBackupVaultNotificationsResponse

See: newGetBackupVaultNotificationsResponse smart constructor.

Constructors

GetBackupVaultNotificationsResponse'

Fields

sNSTopicArn :: Maybe Text

An ARN that uniquely identifies an Amazon Simple Notification Service (Amazon SNS) topic; for example, arn:aws:sns:us-west-2:111122223333:MyTopic.

backupVaultArn :: Maybe Text

An Amazon Resource Name (ARN) that uniquely identifies a backup vault; for example, arn:aws:backup:us-east-1:123456789012:vault:aBackupVault.

backupVaultName :: Maybe Text

The name of a logical container where backups are stored. Backup vaults are identified by names that are unique to the account used to create them and the Region where they are created. They consist of lowercase letters, numbers, and hyphens.

backupVaultEvents :: Maybe [BackupVaultEvent]

An array of events that indicate the status of jobs to back up resources to the backup vault.

httpStatus :: Int

The response's http status code.

newGetBackupVaultNotificationsResponse

Arguments

:: Int
$sel:httpStatus:GetBackupVaultNotificationsResponse'

-> GetBackupVaultNotificationsResponse

Create a value of GetBackupVaultNotificationsResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:sNSTopicArn:GetBackupVaultNotificationsResponse', getBackupVaultNotificationsResponse_sNSTopicArn - An ARN that uniquely identifies an Amazon Simple Notification Service (Amazon SNS) topic; for example, arn:aws:sns:us-west-2:111122223333:MyTopic.

$sel:backupVaultArn:GetBackupVaultNotificationsResponse', getBackupVaultNotificationsResponse_backupVaultArn - An Amazon Resource Name (ARN) that uniquely identifies a backup vault; for example, arn:aws:backup:us-east-1:123456789012:vault:aBackupVault.

$sel:backupVaultName:GetBackupVaultNotifications', getBackupVaultNotificationsResponse_backupVaultName - The name of a logical container where backups are stored. Backup vaults are identified by names that are unique to the account used to create them and the Region where they are created. They consist of lowercase letters, numbers, and hyphens.

$sel:backupVaultEvents:GetBackupVaultNotificationsResponse', getBackupVaultNotificationsResponse_backupVaultEvents - An array of events that indicate the status of jobs to back up resources to the backup vault.

$sel:httpStatus:GetBackupVaultNotificationsResponse', getBackupVaultNotificationsResponse_httpStatus - The response's http status code.

Response Lenses

getBackupVaultNotificationsResponse_sNSTopicArn :: Lens' GetBackupVaultNotificationsResponse (Maybe Text)

An ARN that uniquely identifies an Amazon Simple Notification Service (Amazon SNS) topic; for example, arn:aws:sns:us-west-2:111122223333:MyTopic.

getBackupVaultNotificationsResponse_backupVaultArn :: Lens' GetBackupVaultNotificationsResponse (Maybe Text)

An Amazon Resource Name (ARN) that uniquely identifies a backup vault; for example, arn:aws:backup:us-east-1:123456789012:vault:aBackupVault.

getBackupVaultNotificationsResponse_backupVaultName :: Lens' GetBackupVaultNotificationsResponse (Maybe Text)

The name of a logical container where backups are stored. Backup vaults are identified by names that are unique to the account used to create them and the Region where they are created. They consist of lowercase letters, numbers, and hyphens.

getBackupVaultNotificationsResponse_backupVaultEvents :: Lens' GetBackupVaultNotificationsResponse (Maybe [BackupVaultEvent])

An array of events that indicate the status of jobs to back up resources to the backup vault.

getBackupVaultNotificationsResponse_httpStatus :: Lens' GetBackupVaultNotificationsResponse Int

The response's http status code.