Amazonka.Backup.GetRecoveryPointRestoreMetadata

Description

Returns a set of metadata key-value pairs that were used to create the backup.

Creating a Request

data GetRecoveryPointRestoreMetadata

See: newGetRecoveryPointRestoreMetadata smart constructor.

Constructors

GetRecoveryPointRestoreMetadata'

Fields

backupVaultName :: Text

The name of a logical container where backups are stored. Backup vaults are identified by names that are unique to the account used to create them and the Amazon Web Services Region where they are created. They consist of lowercase letters, numbers, and hyphens.

recoveryPointArn :: Text

An Amazon Resource Name (ARN) that uniquely identifies a recovery point; for example, arn:aws:backup:us-east-1:123456789012:recovery-point:1EB3B5E7-9EB0-435A-A80B-108B488B0D45.

newGetRecoveryPointRestoreMetadata

Arguments

:: Text
$sel:backupVaultName:GetRecoveryPointRestoreMetadata'

-> Text
$sel:recoveryPointArn:GetRecoveryPointRestoreMetadata'

-> GetRecoveryPointRestoreMetadata

Create a value of GetRecoveryPointRestoreMetadata with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:backupVaultName:GetRecoveryPointRestoreMetadata', getRecoveryPointRestoreMetadata_backupVaultName - The name of a logical container where backups are stored. Backup vaults are identified by names that are unique to the account used to create them and the Amazon Web Services Region where they are created. They consist of lowercase letters, numbers, and hyphens.

$sel:recoveryPointArn:GetRecoveryPointRestoreMetadata', getRecoveryPointRestoreMetadata_recoveryPointArn - An Amazon Resource Name (ARN) that uniquely identifies a recovery point; for example, arn:aws:backup:us-east-1:123456789012:recovery-point:1EB3B5E7-9EB0-435A-A80B-108B488B0D45.

Request Lenses

getRecoveryPointRestoreMetadata_backupVaultName :: Lens' GetRecoveryPointRestoreMetadata Text

The name of a logical container where backups are stored. Backup vaults are identified by names that are unique to the account used to create them and the Amazon Web Services Region where they are created. They consist of lowercase letters, numbers, and hyphens.

getRecoveryPointRestoreMetadata_recoveryPointArn :: Lens' GetRecoveryPointRestoreMetadata Text

An Amazon Resource Name (ARN) that uniquely identifies a recovery point; for example, arn:aws:backup:us-east-1:123456789012:recovery-point:1EB3B5E7-9EB0-435A-A80B-108B488B0D45.

Destructuring the Response

data GetRecoveryPointRestoreMetadataResponse

See: newGetRecoveryPointRestoreMetadataResponse smart constructor.

Constructors

GetRecoveryPointRestoreMetadataResponse'

Fields

backupVaultArn :: Maybe Text

An ARN that uniquely identifies a backup vault; for example, arn:aws:backup:us-east-1:123456789012:vault:aBackupVault.

recoveryPointArn :: Maybe Text

An ARN that uniquely identifies a recovery point; for example, arn:aws:backup:us-east-1:123456789012:recovery-point:1EB3B5E7-9EB0-435A-A80B-108B488B0D45.

restoreMetadata :: Maybe (Sensitive (HashMap Text Text))

The set of metadata key-value pairs that describe the original configuration of the backed-up resource. These values vary depending on the service that is being restored.

httpStatus :: Int

The response's http status code.

newGetRecoveryPointRestoreMetadataResponse

Arguments

:: Int
$sel:httpStatus:GetRecoveryPointRestoreMetadataResponse'

-> GetRecoveryPointRestoreMetadataResponse

Create a value of GetRecoveryPointRestoreMetadataResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:backupVaultArn:GetRecoveryPointRestoreMetadataResponse', getRecoveryPointRestoreMetadataResponse_backupVaultArn - An ARN that uniquely identifies a backup vault; for example, arn:aws:backup:us-east-1:123456789012:vault:aBackupVault.

$sel:recoveryPointArn:GetRecoveryPointRestoreMetadata', getRecoveryPointRestoreMetadataResponse_recoveryPointArn - An ARN that uniquely identifies a recovery point; for example, arn:aws:backup:us-east-1:123456789012:recovery-point:1EB3B5E7-9EB0-435A-A80B-108B488B0D45.

$sel:restoreMetadata:GetRecoveryPointRestoreMetadataResponse', getRecoveryPointRestoreMetadataResponse_restoreMetadata - The set of metadata key-value pairs that describe the original configuration of the backed-up resource. These values vary depending on the service that is being restored.

$sel:httpStatus:GetRecoveryPointRestoreMetadataResponse', getRecoveryPointRestoreMetadataResponse_httpStatus - The response's http status code.

Response Lenses

getRecoveryPointRestoreMetadataResponse_backupVaultArn :: Lens' GetRecoveryPointRestoreMetadataResponse (Maybe Text)

An ARN that uniquely identifies a backup vault; for example, arn:aws:backup:us-east-1:123456789012:vault:aBackupVault.

getRecoveryPointRestoreMetadataResponse_recoveryPointArn :: Lens' GetRecoveryPointRestoreMetadataResponse (Maybe Text)

An ARN that uniquely identifies a recovery point; for example, arn:aws:backup:us-east-1:123456789012:recovery-point:1EB3B5E7-9EB0-435A-A80B-108B488B0D45.

getRecoveryPointRestoreMetadataResponse_restoreMetadata :: Lens' GetRecoveryPointRestoreMetadataResponse (Maybe (HashMap Text Text))

The set of metadata key-value pairs that describe the original configuration of the backed-up resource. These values vary depending on the service that is being restored.

getRecoveryPointRestoreMetadataResponse_httpStatus :: Lens' GetRecoveryPointRestoreMetadataResponse Int

The response's http status code.