Amazonka.Backup.ListBackupPlanTemplates

Description

Returns metadata of your saved backup plan templates, including the template ID, name, and the creation and deletion dates.

Creating a Request

data ListBackupPlanTemplates

See: newListBackupPlanTemplates smart constructor.

Constructors

ListBackupPlanTemplates'

Fields

nextToken :: Maybe Text

The next item following a partial list of returned items. For example, if a request is made to return maxResults number of items, NextToken allows you to return more items in your list starting at the location pointed to by the next token.

maxResults :: Maybe Natural

The maximum number of items to be returned.

newListBackupPlanTemplates :: ListBackupPlanTemplates

Create a value of ListBackupPlanTemplates with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:nextToken:ListBackupPlanTemplates', listBackupPlanTemplates_nextToken - The next item following a partial list of returned items. For example, if a request is made to return maxResults number of items, NextToken allows you to return more items in your list starting at the location pointed to by the next token.

$sel:maxResults:ListBackupPlanTemplates', listBackupPlanTemplates_maxResults - The maximum number of items to be returned.

Request Lenses

listBackupPlanTemplates_nextToken :: Lens' ListBackupPlanTemplates (Maybe Text)

The next item following a partial list of returned items. For example, if a request is made to return maxResults number of items, NextToken allows you to return more items in your list starting at the location pointed to by the next token.

listBackupPlanTemplates_maxResults :: Lens' ListBackupPlanTemplates (Maybe Natural)

The maximum number of items to be returned.

Destructuring the Response

data ListBackupPlanTemplatesResponse

See: newListBackupPlanTemplatesResponse smart constructor.

Constructors

ListBackupPlanTemplatesResponse'

Fields

backupPlanTemplatesList :: Maybe [BackupPlanTemplatesListMember]

An array of template list items containing metadata about your saved templates.

nextToken :: Maybe Text

The next item following a partial list of returned items. For example, if a request is made to return maxResults number of items, NextToken allows you to return more items in your list starting at the location pointed to by the next token.

httpStatus :: Int

The response's http status code.

newListBackupPlanTemplatesResponse

Arguments

:: Int
$sel:httpStatus:ListBackupPlanTemplatesResponse'

-> ListBackupPlanTemplatesResponse

Create a value of ListBackupPlanTemplatesResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:backupPlanTemplatesList:ListBackupPlanTemplatesResponse', listBackupPlanTemplatesResponse_backupPlanTemplatesList - An array of template list items containing metadata about your saved templates.

$sel:nextToken:ListBackupPlanTemplates', listBackupPlanTemplatesResponse_nextToken - The next item following a partial list of returned items. For example, if a request is made to return maxResults number of items, NextToken allows you to return more items in your list starting at the location pointed to by the next token.

$sel:httpStatus:ListBackupPlanTemplatesResponse', listBackupPlanTemplatesResponse_httpStatus - The response's http status code.

Response Lenses

listBackupPlanTemplatesResponse_backupPlanTemplatesList :: Lens' ListBackupPlanTemplatesResponse (Maybe [BackupPlanTemplatesListMember])

An array of template list items containing metadata about your saved templates.

listBackupPlanTemplatesResponse_nextToken :: Lens' ListBackupPlanTemplatesResponse (Maybe Text)

The next item following a partial list of returned items. For example, if a request is made to return maxResults number of items, NextToken allows you to return more items in your list starting at the location pointed to by the next token.

listBackupPlanTemplatesResponse_httpStatus :: Lens' ListBackupPlanTemplatesResponse Int

The response's http status code.