Amazonka.Backup.ListBackupSelections

Description

Returns an array containing metadata of the resources associated with the target backup plan.

Creating a Request

data ListBackupSelections

See: newListBackupSelections smart constructor.

Constructors

ListBackupSelections'

Fields

nextToken :: Maybe Text

The next item following a partial list of returned items. For example, if a request is made to return maxResults number of items, NextToken allows you to return more items in your list starting at the location pointed to by the next token.

maxResults :: Maybe Natural

The maximum number of items to be returned.

backupPlanId :: Text

Uniquely identifies a backup plan.

newListBackupSelections

Arguments

:: Text
$sel:backupPlanId:ListBackupSelections'

-> ListBackupSelections

Create a value of ListBackupSelections with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:nextToken:ListBackupSelections', listBackupSelections_nextToken - The next item following a partial list of returned items. For example, if a request is made to return maxResults number of items, NextToken allows you to return more items in your list starting at the location pointed to by the next token.

$sel:maxResults:ListBackupSelections', listBackupSelections_maxResults - The maximum number of items to be returned.

$sel:backupPlanId:ListBackupSelections', listBackupSelections_backupPlanId - Uniquely identifies a backup plan.

Request Lenses

listBackupSelections_nextToken :: Lens' ListBackupSelections (Maybe Text)

The next item following a partial list of returned items. For example, if a request is made to return maxResults number of items, NextToken allows you to return more items in your list starting at the location pointed to by the next token.

listBackupSelections_maxResults :: Lens' ListBackupSelections (Maybe Natural)

The maximum number of items to be returned.

listBackupSelections_backupPlanId :: Lens' ListBackupSelections Text

Uniquely identifies a backup plan.

Destructuring the Response

data ListBackupSelectionsResponse

See: newListBackupSelectionsResponse smart constructor.

Constructors

ListBackupSelectionsResponse'

Fields

nextToken :: Maybe Text

The next item following a partial list of returned items. For example, if a request is made to return maxResults number of items, NextToken allows you to return more items in your list starting at the location pointed to by the next token.

backupSelectionsList :: Maybe [BackupSelectionsListMember]

An array of backup selection list items containing metadata about each resource in the list.

httpStatus :: Int

The response's http status code.

newListBackupSelectionsResponse

Arguments

:: Int
$sel:httpStatus:ListBackupSelectionsResponse'

-> ListBackupSelectionsResponse

Create a value of ListBackupSelectionsResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:nextToken:ListBackupSelections', listBackupSelectionsResponse_nextToken - The next item following a partial list of returned items. For example, if a request is made to return maxResults number of items, NextToken allows you to return more items in your list starting at the location pointed to by the next token.

$sel:backupSelectionsList:ListBackupSelectionsResponse', listBackupSelectionsResponse_backupSelectionsList - An array of backup selection list items containing metadata about each resource in the list.

$sel:httpStatus:ListBackupSelectionsResponse', listBackupSelectionsResponse_httpStatus - The response's http status code.

Response Lenses

listBackupSelectionsResponse_nextToken :: Lens' ListBackupSelectionsResponse (Maybe Text)

The next item following a partial list of returned items. For example, if a request is made to return maxResults number of items, NextToken allows you to return more items in your list starting at the location pointed to by the next token.

listBackupSelectionsResponse_backupSelectionsList :: Lens' ListBackupSelectionsResponse (Maybe [BackupSelectionsListMember])

An array of backup selection list items containing metadata about each resource in the list.

listBackupSelectionsResponse_httpStatus :: Lens' ListBackupSelectionsResponse Int

The response's http status code.