Amazonka.Backup.ListCopyJobs

Description

Returns metadata about your copy jobs.

Creating a Request

data ListCopyJobs

See: newListCopyJobs smart constructor.

Constructors

ListCopyJobs'

Fields

byResourceArn :: Maybe Text

Returns only copy jobs that match the specified resource Amazon Resource Name (ARN).

byCreatedAfter :: Maybe POSIX

Returns only copy jobs that were created after the specified date.

byAccountId :: Maybe Text

The account ID to list the jobs from. Returns only copy jobs associated with the specified account ID.

byCreatedBefore :: Maybe POSIX

Returns only copy jobs that were created before the specified date.

byDestinationVaultArn :: Maybe Text

An Amazon Resource Name (ARN) that uniquely identifies a source backup vault to copy from; for example, arn:aws:backup:us-east-1:123456789012:vault:aBackupVault.

byResourceType :: Maybe Text

Returns only backup jobs for the specified resources:

DynamoDB for Amazon DynamoDB

EBS for Amazon Elastic Block Store

EC2 for Amazon Elastic Compute Cloud

EFS for Amazon Elastic File System

RDS for Amazon Relational Database Service

Aurora for Amazon Aurora

Storage Gateway for Storage Gateway

nextToken :: Maybe Text

The next item following a partial list of returned items. For example, if a request is made to return maxResults number of items, NextToken allows you to return more items in your list starting at the location pointed to by the next token.

byState :: Maybe CopyJobState

Returns only copy jobs that are in the specified state.

maxResults :: Maybe Natural

The maximum number of items to be returned.

newListCopyJobs :: ListCopyJobs

Create a value of ListCopyJobs with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:byResourceArn:ListCopyJobs', listCopyJobs_byResourceArn - Returns only copy jobs that match the specified resource Amazon Resource Name (ARN).

$sel:byCreatedAfter:ListCopyJobs', listCopyJobs_byCreatedAfter - Returns only copy jobs that were created after the specified date.

$sel:byAccountId:ListCopyJobs', listCopyJobs_byAccountId - The account ID to list the jobs from. Returns only copy jobs associated with the specified account ID.

$sel:byCreatedBefore:ListCopyJobs', listCopyJobs_byCreatedBefore - Returns only copy jobs that were created before the specified date.

$sel:byDestinationVaultArn:ListCopyJobs', listCopyJobs_byDestinationVaultArn - An Amazon Resource Name (ARN) that uniquely identifies a source backup vault to copy from; for example, arn:aws:backup:us-east-1:123456789012:vault:aBackupVault.

$sel:byResourceType:ListCopyJobs', listCopyJobs_byResourceType - Returns only backup jobs for the specified resources:

DynamoDB for Amazon DynamoDB

EBS for Amazon Elastic Block Store

EC2 for Amazon Elastic Compute Cloud

EFS for Amazon Elastic File System

RDS for Amazon Relational Database Service

Aurora for Amazon Aurora

Storage Gateway for Storage Gateway

$sel:nextToken:ListCopyJobs', listCopyJobs_nextToken - The next item following a partial list of returned items. For example, if a request is made to return maxResults number of items, NextToken allows you to return more items in your list starting at the location pointed to by the next token.

$sel:byState:ListCopyJobs', listCopyJobs_byState - Returns only copy jobs that are in the specified state.

$sel:maxResults:ListCopyJobs', listCopyJobs_maxResults - The maximum number of items to be returned.

Request Lenses

listCopyJobs_byResourceArn :: Lens' ListCopyJobs (Maybe Text)

Returns only copy jobs that match the specified resource Amazon Resource Name (ARN).

listCopyJobs_byCreatedAfter :: Lens' ListCopyJobs (Maybe UTCTime)

Returns only copy jobs that were created after the specified date.

listCopyJobs_byAccountId :: Lens' ListCopyJobs (Maybe Text)

The account ID to list the jobs from. Returns only copy jobs associated with the specified account ID.

listCopyJobs_byCreatedBefore :: Lens' ListCopyJobs (Maybe UTCTime)

Returns only copy jobs that were created before the specified date.

listCopyJobs_byDestinationVaultArn :: Lens' ListCopyJobs (Maybe Text)

An Amazon Resource Name (ARN) that uniquely identifies a source backup vault to copy from; for example, arn:aws:backup:us-east-1:123456789012:vault:aBackupVault.

listCopyJobs_byResourceType :: Lens' ListCopyJobs (Maybe Text)

Returns only backup jobs for the specified resources:

DynamoDB for Amazon DynamoDB

EBS for Amazon Elastic Block Store

EC2 for Amazon Elastic Compute Cloud

EFS for Amazon Elastic File System

RDS for Amazon Relational Database Service

Aurora for Amazon Aurora

Storage Gateway for Storage Gateway

listCopyJobs_nextToken :: Lens' ListCopyJobs (Maybe Text)

The next item following a partial list of returned items. For example, if a request is made to return maxResults number of items, NextToken allows you to return more items in your list starting at the location pointed to by the next token.

listCopyJobs_byState :: Lens' ListCopyJobs (Maybe CopyJobState)

Returns only copy jobs that are in the specified state.

listCopyJobs_maxResults :: Lens' ListCopyJobs (Maybe Natural)

The maximum number of items to be returned.

Destructuring the Response

data ListCopyJobsResponse

See: newListCopyJobsResponse smart constructor.

Constructors

ListCopyJobsResponse'

Fields

nextToken :: Maybe Text

The next item following a partial list of returned items. For example, if a request is made to return maxResults number of items, NextToken allows you to return more items in your list starting at the location pointed to by the next token.

copyJobs :: Maybe [CopyJob]

An array of structures containing metadata about your copy jobs returned in JSON format.

httpStatus :: Int

The response's http status code.

newListCopyJobsResponse

Arguments

:: Int
$sel:httpStatus:ListCopyJobsResponse'

-> ListCopyJobsResponse

Create a value of ListCopyJobsResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:nextToken:ListCopyJobs', listCopyJobsResponse_nextToken - The next item following a partial list of returned items. For example, if a request is made to return maxResults number of items, NextToken allows you to return more items in your list starting at the location pointed to by the next token.

$sel:copyJobs:ListCopyJobsResponse', listCopyJobsResponse_copyJobs - An array of structures containing metadata about your copy jobs returned in JSON format.

$sel:httpStatus:ListCopyJobsResponse', listCopyJobsResponse_httpStatus - The response's http status code.

Response Lenses

listCopyJobsResponse_nextToken :: Lens' ListCopyJobsResponse (Maybe Text)

The next item following a partial list of returned items. For example, if a request is made to return maxResults number of items, NextToken allows you to return more items in your list starting at the location pointed to by the next token.

listCopyJobsResponse_copyJobs :: Lens' ListCopyJobsResponse (Maybe [CopyJob])

An array of structures containing metadata about your copy jobs returned in JSON format.

listCopyJobsResponse_httpStatus :: Lens' ListCopyJobsResponse Int

The response's http status code.