Amazonka.Backup.StartBackupJob

Description

Starts an on-demand backup job for the specified resource.

Creating a Request

data StartBackupJob

See: newStartBackupJob smart constructor.

Constructors

StartBackupJob'

Fields

idempotencyToken :: Maybe Text

A customer-chosen string that you can use to distinguish between otherwise identical calls to StartBackupJob. Retrying a successful request with the same idempotency token results in a success message with no action taken.

lifecycle :: Maybe Lifecycle

The lifecycle defines when a protected resource is transitioned to cold storage and when it expires. Backup will transition and expire backups automatically according to the lifecycle that you define.

Backups transitioned to cold storage must be stored in cold storage for a minimum of 90 days. Therefore, the “expire after days” setting must be 90 days greater than the “transition to cold after days” setting. The “transition to cold after days” setting cannot be changed after a backup has been transitioned to cold.

Only Amazon EFS file system backups can be transitioned to cold storage.

recoveryPointTags :: Maybe (Sensitive (HashMap Text Text))

To help organize your resources, you can assign your own metadata to the resources that you create. Each tag is a key-value pair.

completeWindowMinutes :: Maybe Integer

A value in minutes during which a successfully started backup must complete, or else AWS Backup will cancel the job. This value is optional. This value begins counting down from when the backup was scheduled. It does not add additional time for StartWindowMinutes, or if the backup started later than scheduled.

backupOptions :: Maybe (HashMap Text Text)

Specifies the backup option for a selected resource. This option is only available for Windows Volume Shadow Copy Service (VSS) backup jobs.

Valid values: Set to "WindowsVSS":"enabled" to enable the WindowsVSS backup option and create a Windows VSS backup. Set to "WindowsVSS""disabled" to create a regular backup. The WindowsVSS option is not enabled by default.

startWindowMinutes :: Maybe Integer

A value in minutes after a backup is scheduled before a job will be canceled if it doesn't start successfully. This value is optional, and the default is 8 hours.

backupVaultName :: Text

The name of a logical container where backups are stored. Backup vaults are identified by names that are unique to the account used to create them and the Amazon Web Services Region where they are created. They consist of lowercase letters, numbers, and hyphens.

resourceArn :: Text

An Amazon Resource Name (ARN) that uniquely identifies a resource. The format of the ARN depends on the resource type.

iamRoleArn :: Text

Specifies the IAM role ARN used to create the target recovery point; for example, arn:aws:iam::123456789012:role/S3Access.

newStartBackupJob

Arguments

:: Text
$sel:backupVaultName:StartBackupJob'

-> Text
$sel:resourceArn:StartBackupJob'

-> Text
$sel:iamRoleArn:StartBackupJob'

-> StartBackupJob

Create a value of StartBackupJob with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:idempotencyToken:StartBackupJob', startBackupJob_idempotencyToken - A customer-chosen string that you can use to distinguish between otherwise identical calls to StartBackupJob. Retrying a successful request with the same idempotency token results in a success message with no action taken.

$sel:lifecycle:StartBackupJob', startBackupJob_lifecycle - The lifecycle defines when a protected resource is transitioned to cold storage and when it expires. Backup will transition and expire backups automatically according to the lifecycle that you define.

Backups transitioned to cold storage must be stored in cold storage for a minimum of 90 days. Therefore, the “expire after days” setting must be 90 days greater than the “transition to cold after days” setting. The “transition to cold after days” setting cannot be changed after a backup has been transitioned to cold.

Only Amazon EFS file system backups can be transitioned to cold storage.

$sel:recoveryPointTags:StartBackupJob', startBackupJob_recoveryPointTags - To help organize your resources, you can assign your own metadata to the resources that you create. Each tag is a key-value pair.

$sel:completeWindowMinutes:StartBackupJob', startBackupJob_completeWindowMinutes - A value in minutes during which a successfully started backup must complete, or else AWS Backup will cancel the job. This value is optional. This value begins counting down from when the backup was scheduled. It does not add additional time for StartWindowMinutes, or if the backup started later than scheduled.

$sel:backupOptions:StartBackupJob', startBackupJob_backupOptions - Specifies the backup option for a selected resource. This option is only available for Windows Volume Shadow Copy Service (VSS) backup jobs.

Valid values: Set to "WindowsVSS":"enabled" to enable the WindowsVSS backup option and create a Windows VSS backup. Set to "WindowsVSS""disabled" to create a regular backup. The WindowsVSS option is not enabled by default.

$sel:startWindowMinutes:StartBackupJob', startBackupJob_startWindowMinutes - A value in minutes after a backup is scheduled before a job will be canceled if it doesn't start successfully. This value is optional, and the default is 8 hours.

$sel:backupVaultName:StartBackupJob', startBackupJob_backupVaultName - The name of a logical container where backups are stored. Backup vaults are identified by names that are unique to the account used to create them and the Amazon Web Services Region where they are created. They consist of lowercase letters, numbers, and hyphens.

$sel:resourceArn:StartBackupJob', startBackupJob_resourceArn - An Amazon Resource Name (ARN) that uniquely identifies a resource. The format of the ARN depends on the resource type.

$sel:iamRoleArn:StartBackupJob', startBackupJob_iamRoleArn - Specifies the IAM role ARN used to create the target recovery point; for example, arn:aws:iam::123456789012:role/S3Access.

Request Lenses

startBackupJob_idempotencyToken :: Lens' StartBackupJob (Maybe Text)

A customer-chosen string that you can use to distinguish between otherwise identical calls to StartBackupJob. Retrying a successful request with the same idempotency token results in a success message with no action taken.

startBackupJob_lifecycle :: Lens' StartBackupJob (Maybe Lifecycle)

The lifecycle defines when a protected resource is transitioned to cold storage and when it expires. Backup will transition and expire backups automatically according to the lifecycle that you define.

Backups transitioned to cold storage must be stored in cold storage for a minimum of 90 days. Therefore, the “expire after days” setting must be 90 days greater than the “transition to cold after days” setting. The “transition to cold after days” setting cannot be changed after a backup has been transitioned to cold.

Only Amazon EFS file system backups can be transitioned to cold storage.

startBackupJob_recoveryPointTags :: Lens' StartBackupJob (Maybe (HashMap Text Text))

To help organize your resources, you can assign your own metadata to the resources that you create. Each tag is a key-value pair.

startBackupJob_completeWindowMinutes :: Lens' StartBackupJob (Maybe Integer)

A value in minutes during which a successfully started backup must complete, or else AWS Backup will cancel the job. This value is optional. This value begins counting down from when the backup was scheduled. It does not add additional time for StartWindowMinutes, or if the backup started later than scheduled.

startBackupJob_backupOptions :: Lens' StartBackupJob (Maybe (HashMap Text Text))

Specifies the backup option for a selected resource. This option is only available for Windows Volume Shadow Copy Service (VSS) backup jobs.

Valid values: Set to "WindowsVSS":"enabled" to enable the WindowsVSS backup option and create a Windows VSS backup. Set to "WindowsVSS""disabled" to create a regular backup. The WindowsVSS option is not enabled by default.

startBackupJob_startWindowMinutes :: Lens' StartBackupJob (Maybe Integer)

A value in minutes after a backup is scheduled before a job will be canceled if it doesn't start successfully. This value is optional, and the default is 8 hours.

startBackupJob_backupVaultName :: Lens' StartBackupJob Text

The name of a logical container where backups are stored. Backup vaults are identified by names that are unique to the account used to create them and the Amazon Web Services Region where they are created. They consist of lowercase letters, numbers, and hyphens.

startBackupJob_resourceArn :: Lens' StartBackupJob Text

An Amazon Resource Name (ARN) that uniquely identifies a resource. The format of the ARN depends on the resource type.

startBackupJob_iamRoleArn :: Lens' StartBackupJob Text

Specifies the IAM role ARN used to create the target recovery point; for example, arn:aws:iam::123456789012:role/S3Access.

Destructuring the Response

data StartBackupJobResponse

See: newStartBackupJobResponse smart constructor.

Constructors

StartBackupJobResponse'

Fields

backupJobId :: Maybe Text

Uniquely identifies a request to Backup to back up a resource.

recoveryPointArn :: Maybe Text

An ARN that uniquely identifies a recovery point; for example, arn:aws:backup:us-east-1:123456789012:recovery-point:1EB3B5E7-9EB0-435A-A80B-108B488B0D45.

creationDate :: Maybe POSIX

The date and time that a backup job is created, in Unix format and Coordinated Universal Time (UTC). The value of CreationDate is accurate to milliseconds. For example, the value 1516925490.087 represents Friday, January 26, 2018 12:11:30.087 AM.

httpStatus :: Int

The response's http status code.

newStartBackupJobResponse

Arguments

:: Int
$sel:httpStatus:StartBackupJobResponse'

-> StartBackupJobResponse

Create a value of StartBackupJobResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:backupJobId:StartBackupJobResponse', startBackupJobResponse_backupJobId - Uniquely identifies a request to Backup to back up a resource.

$sel:recoveryPointArn:StartBackupJobResponse', startBackupJobResponse_recoveryPointArn - An ARN that uniquely identifies a recovery point; for example, arn:aws:backup:us-east-1:123456789012:recovery-point:1EB3B5E7-9EB0-435A-A80B-108B488B0D45.

$sel:creationDate:StartBackupJobResponse', startBackupJobResponse_creationDate - The date and time that a backup job is created, in Unix format and Coordinated Universal Time (UTC). The value of CreationDate is accurate to milliseconds. For example, the value 1516925490.087 represents Friday, January 26, 2018 12:11:30.087 AM.

$sel:httpStatus:StartBackupJobResponse', startBackupJobResponse_httpStatus - The response's http status code.

Response Lenses

startBackupJobResponse_backupJobId :: Lens' StartBackupJobResponse (Maybe Text)

Uniquely identifies a request to Backup to back up a resource.

startBackupJobResponse_recoveryPointArn :: Lens' StartBackupJobResponse (Maybe Text)

An ARN that uniquely identifies a recovery point; for example, arn:aws:backup:us-east-1:123456789012:recovery-point:1EB3B5E7-9EB0-435A-A80B-108B488B0D45.

startBackupJobResponse_creationDate :: Lens' StartBackupJobResponse (Maybe UTCTime)

The date and time that a backup job is created, in Unix format and Coordinated Universal Time (UTC). The value of CreationDate is accurate to milliseconds. For example, the value 1516925490.087 represents Friday, January 26, 2018 12:11:30.087 AM.

startBackupJobResponse_httpStatus :: Lens' StartBackupJobResponse Int

The response's http status code.