Amazonka.Backup.StartRestoreJob

Description

Recovers the saved resource identified by an Amazon Resource Name (ARN).

Creating a Request

data StartRestoreJob

See: newStartRestoreJob smart constructor.

Constructors

StartRestoreJob'

Fields

idempotencyToken :: Maybe Text

A customer-chosen string that you can use to distinguish between otherwise identical calls to StartRestoreJob. Retrying a successful request with the same idempotency token results in a success message with no action taken.

resourceType :: Maybe Text

Starts a job to restore a recovery point for one of the following resources:

DynamoDB for Amazon DynamoDB

EBS for Amazon Elastic Block Store

EC2 for Amazon Elastic Compute Cloud

EFS for Amazon Elastic File System

RDS for Amazon Relational Database Service

Aurora for Amazon Aurora

Storage Gateway for Storage Gateway

recoveryPointArn :: Text

An ARN that uniquely identifies a recovery point; for example, arn:aws:backup:us-east-1:123456789012:recovery-point:1EB3B5E7-9EB0-435A-A80B-108B488B0D45.

metadata :: Sensitive (HashMap Text Text)

A set of metadata key-value pairs. Contains information, such as a resource name, required to restore a recovery point.

You can get configuration metadata about a resource at the time it was backed up by calling GetRecoveryPointRestoreMetadata. However, values in addition to those provided by GetRecoveryPointRestoreMetadata might be required to restore a resource. For example, you might need to provide a new resource name if the original already exists.

You need to specify specific metadata to restore an Amazon Elastic File System (Amazon EFS) instance:

file-system-id: The ID of the Amazon EFS file system that is backed up by Backup. Returned in GetRecoveryPointRestoreMetadata.

Encrypted: A Boolean value that, if true, specifies that the file system is encrypted. If KmsKeyId is specified, Encrypted must be set to true.

KmsKeyId: Specifies the Amazon Web Services KMS key that is used to encrypt the restored file system. You can specify a key from another Amazon Web Services account provided that key it is properly shared with your account via Amazon Web Services KMS.

PerformanceMode: Specifies the throughput mode of the file system.

CreationToken: A user-supplied value that ensures the uniqueness (idempotency) of the request.

newFileSystem: A Boolean value that, if true, specifies that the recovery point is restored to a new Amazon EFS file system.

ItemsToRestore: An array of one to five strings where each string is a file path. Use ItemsToRestore to restore specific files or directories rather than the entire file system. This parameter is optional. For example, "itemsToRestore":"[\"/my.test\"]".

iamRoleArn :: Text

The Amazon Resource Name (ARN) of the IAM role that Backup uses to create the target recovery point; for example, arn:aws:iam::123456789012:role/S3Access.

newStartRestoreJob

Arguments

:: Text
$sel:recoveryPointArn:StartRestoreJob'

-> Text
$sel:iamRoleArn:StartRestoreJob'

-> StartRestoreJob

Create a value of StartRestoreJob with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:idempotencyToken:StartRestoreJob', startRestoreJob_idempotencyToken - A customer-chosen string that you can use to distinguish between otherwise identical calls to StartRestoreJob. Retrying a successful request with the same idempotency token results in a success message with no action taken.

$sel:resourceType:StartRestoreJob', startRestoreJob_resourceType - Starts a job to restore a recovery point for one of the following resources:

DynamoDB for Amazon DynamoDB

EBS for Amazon Elastic Block Store

EC2 for Amazon Elastic Compute Cloud

EFS for Amazon Elastic File System

RDS for Amazon Relational Database Service

Aurora for Amazon Aurora

Storage Gateway for Storage Gateway

$sel:recoveryPointArn:StartRestoreJob', startRestoreJob_recoveryPointArn - An ARN that uniquely identifies a recovery point; for example, arn:aws:backup:us-east-1:123456789012:recovery-point:1EB3B5E7-9EB0-435A-A80B-108B488B0D45.

$sel:metadata:StartRestoreJob', startRestoreJob_metadata - A set of metadata key-value pairs. Contains information, such as a resource name, required to restore a recovery point.

You can get configuration metadata about a resource at the time it was backed up by calling GetRecoveryPointRestoreMetadata. However, values in addition to those provided by GetRecoveryPointRestoreMetadata might be required to restore a resource. For example, you might need to provide a new resource name if the original already exists.

You need to specify specific metadata to restore an Amazon Elastic File System (Amazon EFS) instance:

file-system-id: The ID of the Amazon EFS file system that is backed up by Backup. Returned in GetRecoveryPointRestoreMetadata.

Encrypted: A Boolean value that, if true, specifies that the file system is encrypted. If KmsKeyId is specified, Encrypted must be set to true.

KmsKeyId: Specifies the Amazon Web Services KMS key that is used to encrypt the restored file system. You can specify a key from another Amazon Web Services account provided that key it is properly shared with your account via Amazon Web Services KMS.

PerformanceMode: Specifies the throughput mode of the file system.

CreationToken: A user-supplied value that ensures the uniqueness (idempotency) of the request.

newFileSystem: A Boolean value that, if true, specifies that the recovery point is restored to a new Amazon EFS file system.

ItemsToRestore: An array of one to five strings where each string is a file path. Use ItemsToRestore to restore specific files or directories rather than the entire file system. This parameter is optional. For example, "itemsToRestore":"[\"/my.test\"]".

$sel:iamRoleArn:StartRestoreJob', startRestoreJob_iamRoleArn - The Amazon Resource Name (ARN) of the IAM role that Backup uses to create the target recovery point; for example, arn:aws:iam::123456789012:role/S3Access.

Request Lenses

startRestoreJob_idempotencyToken :: Lens' StartRestoreJob (Maybe Text)

A customer-chosen string that you can use to distinguish between otherwise identical calls to StartRestoreJob. Retrying a successful request with the same idempotency token results in a success message with no action taken.

startRestoreJob_resourceType :: Lens' StartRestoreJob (Maybe Text)

Starts a job to restore a recovery point for one of the following resources:

DynamoDB for Amazon DynamoDB

EBS for Amazon Elastic Block Store

EC2 for Amazon Elastic Compute Cloud

EFS for Amazon Elastic File System

RDS for Amazon Relational Database Service

Aurora for Amazon Aurora

Storage Gateway for Storage Gateway

startRestoreJob_recoveryPointArn :: Lens' StartRestoreJob Text

An ARN that uniquely identifies a recovery point; for example, arn:aws:backup:us-east-1:123456789012:recovery-point:1EB3B5E7-9EB0-435A-A80B-108B488B0D45.

startRestoreJob_metadata :: Lens' StartRestoreJob (HashMap Text Text)

A set of metadata key-value pairs. Contains information, such as a resource name, required to restore a recovery point.

You can get configuration metadata about a resource at the time it was backed up by calling GetRecoveryPointRestoreMetadata. However, values in addition to those provided by GetRecoveryPointRestoreMetadata might be required to restore a resource. For example, you might need to provide a new resource name if the original already exists.

You need to specify specific metadata to restore an Amazon Elastic File System (Amazon EFS) instance:

file-system-id: The ID of the Amazon EFS file system that is backed up by Backup. Returned in GetRecoveryPointRestoreMetadata.

Encrypted: A Boolean value that, if true, specifies that the file system is encrypted. If KmsKeyId is specified, Encrypted must be set to true.

KmsKeyId: Specifies the Amazon Web Services KMS key that is used to encrypt the restored file system. You can specify a key from another Amazon Web Services account provided that key it is properly shared with your account via Amazon Web Services KMS.

PerformanceMode: Specifies the throughput mode of the file system.

CreationToken: A user-supplied value that ensures the uniqueness (idempotency) of the request.

newFileSystem: A Boolean value that, if true, specifies that the recovery point is restored to a new Amazon EFS file system.

ItemsToRestore: An array of one to five strings where each string is a file path. Use ItemsToRestore to restore specific files or directories rather than the entire file system. This parameter is optional. For example, "itemsToRestore":"[\"/my.test\"]".

startRestoreJob_iamRoleArn :: Lens' StartRestoreJob Text

The Amazon Resource Name (ARN) of the IAM role that Backup uses to create the target recovery point; for example, arn:aws:iam::123456789012:role/S3Access.

Destructuring the Response

data StartRestoreJobResponse

See: newStartRestoreJobResponse smart constructor.

Constructors

StartRestoreJobResponse'

Fields

restoreJobId :: Maybe Text

Uniquely identifies the job that restores a recovery point.

httpStatus :: Int

The response's http status code.

newStartRestoreJobResponse

Arguments

:: Int
$sel:httpStatus:StartRestoreJobResponse'

-> StartRestoreJobResponse

Create a value of StartRestoreJobResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:restoreJobId:StartRestoreJobResponse', startRestoreJobResponse_restoreJobId - Uniquely identifies the job that restores a recovery point.

$sel:httpStatus:StartRestoreJobResponse', startRestoreJobResponse_httpStatus - The response's http status code.

Response Lenses

startRestoreJobResponse_restoreJobId :: Lens' StartRestoreJobResponse (Maybe Text)

Uniquely identifies the job that restores a recovery point.

startRestoreJobResponse_httpStatus :: Lens' StartRestoreJobResponse Int

The response's http status code.