Amazonka.Backup.Types.AdvancedBackupSetting

Description

Documentation

data AdvancedBackupSetting

A list of backup options for each resource type.

See: newAdvancedBackupSetting smart constructor.

Constructors

AdvancedBackupSetting'

Fields

resourceType :: Maybe Text

Specifies an object containing resource type and backup options. The only supported resource type is Amazon EC2 instances with Windows Volume Shadow Copy Service (VSS). For a CloudFormation example, see the sample CloudFormation template to enable Windows VSS in the Backup User Guide.

Valid values: EC2.

backupOptions :: Maybe (HashMap Text Text)

Specifies the backup option for a selected resource. This option is only available for Windows VSS backup jobs.

Valid values:

Set to "WindowsVSS":"enabled" to enable the WindowsVSS backup option and create a Windows VSS backup.

Set to "WindowsVSS":"disabled" to create a regular backup. The WindowsVSS option is not enabled by default.

If you specify an invalid option, you get an InvalidParameterValueException exception.

For more information about Windows VSS backups, see Creating a VSS-Enabled Windows Backup.

newAdvancedBackupSetting :: AdvancedBackupSetting

Create a value of AdvancedBackupSetting with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:resourceType:AdvancedBackupSetting', advancedBackupSetting_resourceType - Specifies an object containing resource type and backup options. The only supported resource type is Amazon EC2 instances with Windows Volume Shadow Copy Service (VSS). For a CloudFormation example, see the sample CloudFormation template to enable Windows VSS in the Backup User Guide.

Valid values: EC2.

$sel:backupOptions:AdvancedBackupSetting', advancedBackupSetting_backupOptions - Specifies the backup option for a selected resource. This option is only available for Windows VSS backup jobs.

Valid values:

Set to "WindowsVSS":"enabled" to enable the WindowsVSS backup option and create a Windows VSS backup.

Set to "WindowsVSS":"disabled" to create a regular backup. The WindowsVSS option is not enabled by default.

If you specify an invalid option, you get an InvalidParameterValueException exception.

For more information about Windows VSS backups, see Creating a VSS-Enabled Windows Backup.

advancedBackupSetting_resourceType :: Lens' AdvancedBackupSetting (Maybe Text)

Specifies an object containing resource type and backup options. The only supported resource type is Amazon EC2 instances with Windows Volume Shadow Copy Service (VSS). For a CloudFormation example, see the sample CloudFormation template to enable Windows VSS in the Backup User Guide.

Valid values: EC2.

advancedBackupSetting_backupOptions :: Lens' AdvancedBackupSetting (Maybe (HashMap Text Text))

Specifies the backup option for a selected resource. This option is only available for Windows VSS backup jobs.

Valid values:

Set to "WindowsVSS":"enabled" to enable the WindowsVSS backup option and create a Windows VSS backup.

Set to "WindowsVSS":"disabled" to create a regular backup. The WindowsVSS option is not enabled by default.

If you specify an invalid option, you get an InvalidParameterValueException exception.

For more information about Windows VSS backups, see Creating a VSS-Enabled Windows Backup.