Amazonka.Backup.Types.BackupJob

Description

Documentation

data BackupJob

Contains detailed information about a backup job.

See: newBackupJob smart constructor.

Constructors

BackupJob'

Fields

iamRoleArn :: Maybe Text

Specifies the IAM role ARN used to create the target recovery point. IAM roles other than the default role must include either AWSBackup or AwsBackup in the role name. For example, arn:aws:iam::123456789012:role/AWSBackupRDSAccess. Role names without those strings lack permissions to perform backup jobs.

state :: Maybe BackupJobState

The current state of a resource recovery point.

resourceType :: Maybe Text

The type of Amazon Web Services resource to be backed up; for example, an Amazon Elastic Block Store (Amazon EBS) volume or an Amazon Relational Database Service (Amazon RDS) database. For Windows Volume Shadow Copy Service (VSS) backups, the only supported resource type is Amazon EC2.

percentDone :: Maybe Text

Contains an estimated percentage complete of a job at the time the job status was queried.

startBy :: Maybe POSIX

Specifies the time in Unix format and Coordinated Universal Time (UTC) when a backup job must be started before it is canceled. The value is calculated by adding the start window to the scheduled time. So if the scheduled time were 6:00 PM and the start window is 2 hours, the StartBy time would be 8:00 PM on the date specified. The value of StartBy is accurate to milliseconds. For example, the value 1516925490.087 represents Friday, January 26, 2018 12:11:30.087 AM.

createdBy :: Maybe RecoveryPointCreator

Contains identifying information about the creation of a backup job, including the BackupPlanArn, BackupPlanId, BackupPlanVersion, and BackupRuleId of the backup plan used to create it.

expectedCompletionDate :: Maybe POSIX

The date and time a job to back up resources is expected to be completed, in Unix format and Coordinated Universal Time (UTC). The value of ExpectedCompletionDate is accurate to milliseconds. For example, the value 1516925490.087 represents Friday, January 26, 2018 12:11:30.087 AM.

bytesTransferred :: Maybe Integer

The size in bytes transferred to a backup vault at the time that the job status was queried.

backupVaultArn :: Maybe Text

An Amazon Resource Name (ARN) that uniquely identifies a backup vault; for example, arn:aws:backup:us-east-1:123456789012:vault:aBackupVault.

accountId :: Maybe Text

The account ID that owns the backup job.

backupJobId :: Maybe Text

Uniquely identifies a request to Backup to back up a resource.

resourceArn :: Maybe Text

An ARN that uniquely identifies a resource. The format of the ARN depends on the resource type.

statusMessage :: Maybe Text

A detailed message explaining the status of the job to back up a resource.

recoveryPointArn :: Maybe Text

An ARN that uniquely identifies a recovery point; for example, arn:aws:backup:us-east-1:123456789012:recovery-point:1EB3B5E7-9EB0-435A-A80B-108B488B0D45.

backupSizeInBytes :: Maybe Integer

The size, in bytes, of a backup.

creationDate :: Maybe POSIX

The date and time a backup job is created, in Unix format and Coordinated Universal Time (UTC). The value of CreationDate is accurate to milliseconds. For example, the value 1516925490.087 represents Friday, January 26, 2018 12:11:30.087 AM.

completionDate :: Maybe POSIX

The date and time a job to create a backup job is completed, in Unix format and Coordinated Universal Time (UTC). The value of CompletionDate is accurate to milliseconds. For example, the value 1516925490.087 represents Friday, January 26, 2018 12:11:30.087 AM.

backupVaultName :: Maybe Text

The name of a logical container where backups are stored. Backup vaults are identified by names that are unique to the account used to create them and the Amazon Web Services Region where they are created. They consist of lowercase letters, numbers, and hyphens.

backupType :: Maybe Text

Represents the type of backup for a backup job.

backupOptions :: Maybe (HashMap Text Text)

Specifies the backup option for a selected resource. This option is only available for Windows Volume Shadow Copy Service (VSS) backup jobs.

Valid values: Set to "WindowsVSS":"enabled" to enable the WindowsVSS backup option and create a Windows VSS backup. Set to "WindowsVSS":"disabled" to create a regular backup. If you specify an invalid option, you get an InvalidParameterValueException exception.

newBackupJob :: BackupJob

Create a value of BackupJob with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:iamRoleArn:BackupJob', backupJob_iamRoleArn - Specifies the IAM role ARN used to create the target recovery point. IAM roles other than the default role must include either AWSBackup or AwsBackup in the role name. For example, arn:aws:iam::123456789012:role/AWSBackupRDSAccess. Role names without those strings lack permissions to perform backup jobs.

$sel:state:BackupJob', backupJob_state - The current state of a resource recovery point.

$sel:resourceType:BackupJob', backupJob_resourceType - The type of Amazon Web Services resource to be backed up; for example, an Amazon Elastic Block Store (Amazon EBS) volume or an Amazon Relational Database Service (Amazon RDS) database. For Windows Volume Shadow Copy Service (VSS) backups, the only supported resource type is Amazon EC2.

$sel:percentDone:BackupJob', backupJob_percentDone - Contains an estimated percentage complete of a job at the time the job status was queried.

$sel:startBy:BackupJob', backupJob_startBy - Specifies the time in Unix format and Coordinated Universal Time (UTC) when a backup job must be started before it is canceled. The value is calculated by adding the start window to the scheduled time. So if the scheduled time were 6:00 PM and the start window is 2 hours, the StartBy time would be 8:00 PM on the date specified. The value of StartBy is accurate to milliseconds. For example, the value 1516925490.087 represents Friday, January 26, 2018 12:11:30.087 AM.

$sel:createdBy:BackupJob', backupJob_createdBy - Contains identifying information about the creation of a backup job, including the BackupPlanArn, BackupPlanId, BackupPlanVersion, and BackupRuleId of the backup plan used to create it.

$sel:expectedCompletionDate:BackupJob', backupJob_expectedCompletionDate - The date and time a job to back up resources is expected to be completed, in Unix format and Coordinated Universal Time (UTC). The value of ExpectedCompletionDate is accurate to milliseconds. For example, the value 1516925490.087 represents Friday, January 26, 2018 12:11:30.087 AM.

$sel:bytesTransferred:BackupJob', backupJob_bytesTransferred - The size in bytes transferred to a backup vault at the time that the job status was queried.

$sel:backupVaultArn:BackupJob', backupJob_backupVaultArn - An Amazon Resource Name (ARN) that uniquely identifies a backup vault; for example, arn:aws:backup:us-east-1:123456789012:vault:aBackupVault.

$sel:accountId:BackupJob', backupJob_accountId - The account ID that owns the backup job.

$sel:backupJobId:BackupJob', backupJob_backupJobId - Uniquely identifies a request to Backup to back up a resource.

$sel:resourceArn:BackupJob', backupJob_resourceArn - An ARN that uniquely identifies a resource. The format of the ARN depends on the resource type.

$sel:statusMessage:BackupJob', backupJob_statusMessage - A detailed message explaining the status of the job to back up a resource.

$sel:recoveryPointArn:BackupJob', backupJob_recoveryPointArn - An ARN that uniquely identifies a recovery point; for example, arn:aws:backup:us-east-1:123456789012:recovery-point:1EB3B5E7-9EB0-435A-A80B-108B488B0D45.

$sel:backupSizeInBytes:BackupJob', backupJob_backupSizeInBytes - The size, in bytes, of a backup.

$sel:creationDate:BackupJob', backupJob_creationDate - The date and time a backup job is created, in Unix format and Coordinated Universal Time (UTC). The value of CreationDate is accurate to milliseconds. For example, the value 1516925490.087 represents Friday, January 26, 2018 12:11:30.087 AM.

$sel:completionDate:BackupJob', backupJob_completionDate - The date and time a job to create a backup job is completed, in Unix format and Coordinated Universal Time (UTC). The value of CompletionDate is accurate to milliseconds. For example, the value 1516925490.087 represents Friday, January 26, 2018 12:11:30.087 AM.

$sel:backupVaultName:BackupJob', backupJob_backupVaultName - The name of a logical container where backups are stored. Backup vaults are identified by names that are unique to the account used to create them and the Amazon Web Services Region where they are created. They consist of lowercase letters, numbers, and hyphens.

$sel:backupType:BackupJob', backupJob_backupType - Represents the type of backup for a backup job.

$sel:backupOptions:BackupJob', backupJob_backupOptions - Specifies the backup option for a selected resource. This option is only available for Windows Volume Shadow Copy Service (VSS) backup jobs.

Valid values: Set to "WindowsVSS":"enabled" to enable the WindowsVSS backup option and create a Windows VSS backup. Set to "WindowsVSS":"disabled" to create a regular backup. If you specify an invalid option, you get an InvalidParameterValueException exception.

backupJob_iamRoleArn :: Lens' BackupJob (Maybe Text)

Specifies the IAM role ARN used to create the target recovery point. IAM roles other than the default role must include either AWSBackup or AwsBackup in the role name. For example, arn:aws:iam::123456789012:role/AWSBackupRDSAccess. Role names without those strings lack permissions to perform backup jobs.

backupJob_state :: Lens' BackupJob (Maybe BackupJobState)

The current state of a resource recovery point.

backupJob_resourceType :: Lens' BackupJob (Maybe Text)

The type of Amazon Web Services resource to be backed up; for example, an Amazon Elastic Block Store (Amazon EBS) volume or an Amazon Relational Database Service (Amazon RDS) database. For Windows Volume Shadow Copy Service (VSS) backups, the only supported resource type is Amazon EC2.

backupJob_percentDone :: Lens' BackupJob (Maybe Text)

Contains an estimated percentage complete of a job at the time the job status was queried.

backupJob_startBy :: Lens' BackupJob (Maybe UTCTime)

Specifies the time in Unix format and Coordinated Universal Time (UTC) when a backup job must be started before it is canceled. The value is calculated by adding the start window to the scheduled time. So if the scheduled time were 6:00 PM and the start window is 2 hours, the StartBy time would be 8:00 PM on the date specified. The value of StartBy is accurate to milliseconds. For example, the value 1516925490.087 represents Friday, January 26, 2018 12:11:30.087 AM.

backupJob_createdBy :: Lens' BackupJob (Maybe RecoveryPointCreator)

Contains identifying information about the creation of a backup job, including the BackupPlanArn, BackupPlanId, BackupPlanVersion, and BackupRuleId of the backup plan used to create it.

backupJob_expectedCompletionDate :: Lens' BackupJob (Maybe UTCTime)

The date and time a job to back up resources is expected to be completed, in Unix format and Coordinated Universal Time (UTC). The value of ExpectedCompletionDate is accurate to milliseconds. For example, the value 1516925490.087 represents Friday, January 26, 2018 12:11:30.087 AM.

backupJob_bytesTransferred :: Lens' BackupJob (Maybe Integer)

The size in bytes transferred to a backup vault at the time that the job status was queried.

backupJob_backupVaultArn :: Lens' BackupJob (Maybe Text)

An Amazon Resource Name (ARN) that uniquely identifies a backup vault; for example, arn:aws:backup:us-east-1:123456789012:vault:aBackupVault.

backupJob_accountId :: Lens' BackupJob (Maybe Text)

The account ID that owns the backup job.

backupJob_backupJobId :: Lens' BackupJob (Maybe Text)

Uniquely identifies a request to Backup to back up a resource.

backupJob_resourceArn :: Lens' BackupJob (Maybe Text)

An ARN that uniquely identifies a resource. The format of the ARN depends on the resource type.

backupJob_statusMessage :: Lens' BackupJob (Maybe Text)

A detailed message explaining the status of the job to back up a resource.

backupJob_recoveryPointArn :: Lens' BackupJob (Maybe Text)

An ARN that uniquely identifies a recovery point; for example, arn:aws:backup:us-east-1:123456789012:recovery-point:1EB3B5E7-9EB0-435A-A80B-108B488B0D45.

backupJob_backupSizeInBytes :: Lens' BackupJob (Maybe Integer)

The size, in bytes, of a backup.

backupJob_creationDate :: Lens' BackupJob (Maybe UTCTime)

The date and time a backup job is created, in Unix format and Coordinated Universal Time (UTC). The value of CreationDate is accurate to milliseconds. For example, the value 1516925490.087 represents Friday, January 26, 2018 12:11:30.087 AM.

backupJob_completionDate :: Lens' BackupJob (Maybe UTCTime)

The date and time a job to create a backup job is completed, in Unix format and Coordinated Universal Time (UTC). The value of CompletionDate is accurate to milliseconds. For example, the value 1516925490.087 represents Friday, January 26, 2018 12:11:30.087 AM.

backupJob_backupVaultName :: Lens' BackupJob (Maybe Text)

The name of a logical container where backups are stored. Backup vaults are identified by names that are unique to the account used to create them and the Amazon Web Services Region where they are created. They consist of lowercase letters, numbers, and hyphens.

backupJob_backupType :: Lens' BackupJob (Maybe Text)

Represents the type of backup for a backup job.

backupJob_backupOptions :: Lens' BackupJob (Maybe (HashMap Text Text))

Specifies the backup option for a selected resource. This option is only available for Windows Volume Shadow Copy Service (VSS) backup jobs.

Valid values: Set to "WindowsVSS":"enabled" to enable the WindowsVSS backup option and create a Windows VSS backup. Set to "WindowsVSS":"disabled" to create a regular backup. If you specify an invalid option, you get an InvalidParameterValueException exception.