Amazonka.Backup.Types.BackupPlanInput

Description

Documentation

data BackupPlanInput

Contains an optional backup plan display name and an array of BackupRule objects, each of which specifies a backup rule. Each rule in a backup plan is a separate scheduled task.

See: newBackupPlanInput smart constructor.

Constructors

BackupPlanInput'

Fields

advancedBackupSettings :: Maybe [AdvancedBackupSetting]

Specifies a list of BackupOptions for each resource type. These settings are only available for Windows Volume Shadow Copy Service (VSS) backup jobs.

backupPlanName :: Text

The optional display name of a backup plan.

rules :: [BackupRuleInput]

An array of BackupRule objects, each of which specifies a scheduled task that is used to back up a selection of resources.

newBackupPlanInput

Arguments

:: Text
$sel:backupPlanName:BackupPlanInput'

-> BackupPlanInput

Create a value of BackupPlanInput with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:advancedBackupSettings:BackupPlanInput', backupPlanInput_advancedBackupSettings - Specifies a list of BackupOptions for each resource type. These settings are only available for Windows Volume Shadow Copy Service (VSS) backup jobs.

$sel:backupPlanName:BackupPlanInput', backupPlanInput_backupPlanName - The optional display name of a backup plan.

$sel:rules:BackupPlanInput', backupPlanInput_rules - An array of BackupRule objects, each of which specifies a scheduled task that is used to back up a selection of resources.

backupPlanInput_advancedBackupSettings :: Lens' BackupPlanInput (Maybe [AdvancedBackupSetting])

Specifies a list of BackupOptions for each resource type. These settings are only available for Windows Volume Shadow Copy Service (VSS) backup jobs.

backupPlanInput_backupPlanName :: Lens' BackupPlanInput Text

The optional display name of a backup plan.

backupPlanInput_rules :: Lens' BackupPlanInput [BackupRuleInput]

An array of BackupRule objects, each of which specifies a scheduled task that is used to back up a selection of resources.