Amazonka.Backup.Types.BackupPlansListMember

Description

Documentation

data BackupPlansListMember

Contains metadata about a backup plan.

See: newBackupPlansListMember smart constructor.

Constructors

BackupPlansListMember'

Fields

versionId :: Maybe Text

Unique, randomly generated, Unicode, UTF-8 encoded strings that are at most 1,024 bytes long. Version IDs cannot be edited.

backupPlanName :: Maybe Text

The display name of a saved backup plan.

advancedBackupSettings :: Maybe [AdvancedBackupSetting]

Contains a list of BackupOptions for a resource type.

backupPlanId :: Maybe Text

Uniquely identifies a backup plan.

creatorRequestId :: Maybe Text

A unique string that identifies the request and allows failed requests to be retried without the risk of running the operation twice.

backupPlanArn :: Maybe Text

An Amazon Resource Name (ARN) that uniquely identifies a backup plan; for example, arn:aws:backup:us-east-1:123456789012:plan:8F81F553-3A74-4A3F-B93D-B3360DC80C50.

lastExecutionDate :: Maybe POSIX

The last time a job to back up resources was run with this rule. A date and time, in Unix format and Coordinated Universal Time (UTC). The value of LastExecutionDate is accurate to milliseconds. For example, the value 1516925490.087 represents Friday, January 26, 2018 12:11:30.087 AM.

creationDate :: Maybe POSIX

The date and time a resource backup plan is created, in Unix format and Coordinated Universal Time (UTC). The value of CreationDate is accurate to milliseconds. For example, the value 1516925490.087 represents Friday, January 26, 2018 12:11:30.087 AM.

deletionDate :: Maybe POSIX

The date and time a backup plan is deleted, in Unix format and Coordinated Universal Time (UTC). The value of DeletionDate is accurate to milliseconds. For example, the value 1516925490.087 represents Friday, January 26, 2018 12:11:30.087 AM.

newBackupPlansListMember :: BackupPlansListMember

Create a value of BackupPlansListMember with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:versionId:BackupPlansListMember', backupPlansListMember_versionId - Unique, randomly generated, Unicode, UTF-8 encoded strings that are at most 1,024 bytes long. Version IDs cannot be edited.

$sel:backupPlanName:BackupPlansListMember', backupPlansListMember_backupPlanName - The display name of a saved backup plan.

$sel:advancedBackupSettings:BackupPlansListMember', backupPlansListMember_advancedBackupSettings - Contains a list of BackupOptions for a resource type.

$sel:backupPlanId:BackupPlansListMember', backupPlansListMember_backupPlanId - Uniquely identifies a backup plan.

$sel:creatorRequestId:BackupPlansListMember', backupPlansListMember_creatorRequestId - A unique string that identifies the request and allows failed requests to be retried without the risk of running the operation twice.

$sel:backupPlanArn:BackupPlansListMember', backupPlansListMember_backupPlanArn - An Amazon Resource Name (ARN) that uniquely identifies a backup plan; for example, arn:aws:backup:us-east-1:123456789012:plan:8F81F553-3A74-4A3F-B93D-B3360DC80C50.

$sel:lastExecutionDate:BackupPlansListMember', backupPlansListMember_lastExecutionDate - The last time a job to back up resources was run with this rule. A date and time, in Unix format and Coordinated Universal Time (UTC). The value of LastExecutionDate is accurate to milliseconds. For example, the value 1516925490.087 represents Friday, January 26, 2018 12:11:30.087 AM.

$sel:creationDate:BackupPlansListMember', backupPlansListMember_creationDate - The date and time a resource backup plan is created, in Unix format and Coordinated Universal Time (UTC). The value of CreationDate is accurate to milliseconds. For example, the value 1516925490.087 represents Friday, January 26, 2018 12:11:30.087 AM.

$sel:deletionDate:BackupPlansListMember', backupPlansListMember_deletionDate - The date and time a backup plan is deleted, in Unix format and Coordinated Universal Time (UTC). The value of DeletionDate is accurate to milliseconds. For example, the value 1516925490.087 represents Friday, January 26, 2018 12:11:30.087 AM.

backupPlansListMember_versionId :: Lens' BackupPlansListMember (Maybe Text)

Unique, randomly generated, Unicode, UTF-8 encoded strings that are at most 1,024 bytes long. Version IDs cannot be edited.

backupPlansListMember_backupPlanName :: Lens' BackupPlansListMember (Maybe Text)

The display name of a saved backup plan.

backupPlansListMember_advancedBackupSettings :: Lens' BackupPlansListMember (Maybe [AdvancedBackupSetting])

Contains a list of BackupOptions for a resource type.

backupPlansListMember_backupPlanId :: Lens' BackupPlansListMember (Maybe Text)

Uniquely identifies a backup plan.

backupPlansListMember_creatorRequestId :: Lens' BackupPlansListMember (Maybe Text)

A unique string that identifies the request and allows failed requests to be retried without the risk of running the operation twice.

backupPlansListMember_backupPlanArn :: Lens' BackupPlansListMember (Maybe Text)

An Amazon Resource Name (ARN) that uniquely identifies a backup plan; for example, arn:aws:backup:us-east-1:123456789012:plan:8F81F553-3A74-4A3F-B93D-B3360DC80C50.

backupPlansListMember_lastExecutionDate :: Lens' BackupPlansListMember (Maybe UTCTime)

The last time a job to back up resources was run with this rule. A date and time, in Unix format and Coordinated Universal Time (UTC). The value of LastExecutionDate is accurate to milliseconds. For example, the value 1516925490.087 represents Friday, January 26, 2018 12:11:30.087 AM.

backupPlansListMember_creationDate :: Lens' BackupPlansListMember (Maybe UTCTime)

The date and time a resource backup plan is created, in Unix format and Coordinated Universal Time (UTC). The value of CreationDate is accurate to milliseconds. For example, the value 1516925490.087 represents Friday, January 26, 2018 12:11:30.087 AM.

backupPlansListMember_deletionDate :: Lens' BackupPlansListMember (Maybe UTCTime)

The date and time a backup plan is deleted, in Unix format and Coordinated Universal Time (UTC). The value of DeletionDate is accurate to milliseconds. For example, the value 1516925490.087 represents Friday, January 26, 2018 12:11:30.087 AM.