Amazonka.Backup.Types.BackupSelection

Description

Documentation

data BackupSelection

Used to specify a set of resources to a backup plan.

See: newBackupSelection smart constructor.

Constructors

BackupSelection'

Fields

resources :: Maybe [Text]

An array of strings that contain Amazon Resource Names (ARNs) of resources to assign to a backup plan.

listOfTags :: Maybe [Condition]

An array of conditions used to specify a set of resources to assign to a backup plan; for example, "StringEquals": {"ec2:ResourceTag/Department": "accounting". Assigns the backup plan to every resource with at least one matching tag.

selectionName :: Text

The display name of a resource selection document.

iamRoleArn :: Text

The ARN of the IAM role that Backup uses to authenticate when backing up the target resource; for example, arn:aws:iam::123456789012:role/S3Access.

newBackupSelection

Arguments

:: Text
$sel:selectionName:BackupSelection'

-> Text
$sel:iamRoleArn:BackupSelection'

-> BackupSelection

Create a value of BackupSelection with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:resources:BackupSelection', backupSelection_resources - An array of strings that contain Amazon Resource Names (ARNs) of resources to assign to a backup plan.

$sel:listOfTags:BackupSelection', backupSelection_listOfTags - An array of conditions used to specify a set of resources to assign to a backup plan; for example, "StringEquals": {"ec2:ResourceTag/Department": "accounting". Assigns the backup plan to every resource with at least one matching tag.

$sel:selectionName:BackupSelection', backupSelection_selectionName - The display name of a resource selection document.

$sel:iamRoleArn:BackupSelection', backupSelection_iamRoleArn - The ARN of the IAM role that Backup uses to authenticate when backing up the target resource; for example, arn:aws:iam::123456789012:role/S3Access.

backupSelection_resources :: Lens' BackupSelection (Maybe [Text])

An array of strings that contain Amazon Resource Names (ARNs) of resources to assign to a backup plan.

backupSelection_listOfTags :: Lens' BackupSelection (Maybe [Condition])

An array of conditions used to specify a set of resources to assign to a backup plan; for example, "StringEquals": {"ec2:ResourceTag/Department": "accounting". Assigns the backup plan to every resource with at least one matching tag.

backupSelection_selectionName :: Lens' BackupSelection Text

The display name of a resource selection document.

backupSelection_iamRoleArn :: Lens' BackupSelection Text

The ARN of the IAM role that Backup uses to authenticate when backing up the target resource; for example, arn:aws:iam::123456789012:role/S3Access.