Amazonka.Backup.Types.BackupSelectionsListMember

Description

Documentation

data BackupSelectionsListMember

Contains metadata about a BackupSelection object.

See: newBackupSelectionsListMember smart constructor.

Constructors

BackupSelectionsListMember'

Fields

iamRoleArn :: Maybe Text

Specifies the IAM role Amazon Resource Name (ARN) to create the target recovery point; for example, arn:aws:iam::123456789012:role/S3Access.

selectionName :: Maybe Text

The display name of a resource selection document.

selectionId :: Maybe Text

Uniquely identifies a request to assign a set of resources to a backup plan.

backupPlanId :: Maybe Text

Uniquely identifies a backup plan.

creatorRequestId :: Maybe Text

A unique string that identifies the request and allows failed requests to be retried without the risk of running the operation twice.

creationDate :: Maybe POSIX

The date and time a backup plan is created, in Unix format and Coordinated Universal Time (UTC). The value of CreationDate is accurate to milliseconds. For example, the value 1516925490.087 represents Friday, January 26, 2018 12:11:30.087 AM.

newBackupSelectionsListMember :: BackupSelectionsListMember

Create a value of BackupSelectionsListMember with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:iamRoleArn:BackupSelectionsListMember', backupSelectionsListMember_iamRoleArn - Specifies the IAM role Amazon Resource Name (ARN) to create the target recovery point; for example, arn:aws:iam::123456789012:role/S3Access.

$sel:selectionName:BackupSelectionsListMember', backupSelectionsListMember_selectionName - The display name of a resource selection document.

$sel:selectionId:BackupSelectionsListMember', backupSelectionsListMember_selectionId - Uniquely identifies a request to assign a set of resources to a backup plan.

$sel:backupPlanId:BackupSelectionsListMember', backupSelectionsListMember_backupPlanId - Uniquely identifies a backup plan.

$sel:creatorRequestId:BackupSelectionsListMember', backupSelectionsListMember_creatorRequestId - A unique string that identifies the request and allows failed requests to be retried without the risk of running the operation twice.

$sel:creationDate:BackupSelectionsListMember', backupSelectionsListMember_creationDate - The date and time a backup plan is created, in Unix format and Coordinated Universal Time (UTC). The value of CreationDate is accurate to milliseconds. For example, the value 1516925490.087 represents Friday, January 26, 2018 12:11:30.087 AM.

backupSelectionsListMember_iamRoleArn :: Lens' BackupSelectionsListMember (Maybe Text)

Specifies the IAM role Amazon Resource Name (ARN) to create the target recovery point; for example, arn:aws:iam::123456789012:role/S3Access.

backupSelectionsListMember_selectionName :: Lens' BackupSelectionsListMember (Maybe Text)

The display name of a resource selection document.

backupSelectionsListMember_selectionId :: Lens' BackupSelectionsListMember (Maybe Text)

Uniquely identifies a request to assign a set of resources to a backup plan.

backupSelectionsListMember_backupPlanId :: Lens' BackupSelectionsListMember (Maybe Text)

Uniquely identifies a backup plan.

backupSelectionsListMember_creatorRequestId :: Lens' BackupSelectionsListMember (Maybe Text)

A unique string that identifies the request and allows failed requests to be retried without the risk of running the operation twice.

backupSelectionsListMember_creationDate :: Lens' BackupSelectionsListMember (Maybe UTCTime)

The date and time a backup plan is created, in Unix format and Coordinated Universal Time (UTC). The value of CreationDate is accurate to milliseconds. For example, the value 1516925490.087 represents Friday, January 26, 2018 12:11:30.087 AM.