Amazonka.Backup.Types.Condition

Description

Documentation

data Condition

Contains an array of triplets made up of a condition type (such as StringEquals), a key, and a value. Conditions are used to filter resources in a selection that is assigned to a backup plan.

See: newCondition smart constructor.

Constructors

Condition'

Fields

conditionType :: ConditionType

An operation, such as StringEquals, that is applied to a key-value pair used to filter resources in a selection.

conditionKey :: Text

The key in a key-value pair. For example, in "ec2:ResourceTag/Department": "accounting", "ec2:ResourceTag/Department" is the key.

conditionValue :: Text

The value in a key-value pair. For example, in "ec2:ResourceTag/Department": "accounting", "accounting" is the value.

newCondition

Arguments

:: ConditionType
$sel:conditionType:Condition'

-> Text
$sel:conditionKey:Condition'

-> Text
$sel:conditionValue:Condition'

-> Condition

Create a value of Condition with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:conditionType:Condition', condition_conditionType - An operation, such as StringEquals, that is applied to a key-value pair used to filter resources in a selection.

$sel:conditionKey:Condition', condition_conditionKey - The key in a key-value pair. For example, in "ec2:ResourceTag/Department": "accounting", "ec2:ResourceTag/Department" is the key.

$sel:conditionValue:Condition', condition_conditionValue - The value in a key-value pair. For example, in "ec2:ResourceTag/Department": "accounting", "accounting" is the value.

condition_conditionType :: Lens' Condition ConditionType

An operation, such as StringEquals, that is applied to a key-value pair used to filter resources in a selection.

condition_conditionKey :: Lens' Condition Text

The key in a key-value pair. For example, in "ec2:ResourceTag/Department": "accounting", "ec2:ResourceTag/Department" is the key.

condition_conditionValue :: Lens' Condition Text

The value in a key-value pair. For example, in "ec2:ResourceTag/Department": "accounting", "accounting" is the value.