Amazonka.Backup.Types.CopyJob

Description

Documentation

data CopyJob

Contains detailed information about a copy job.

See: newCopyJob smart constructor.

Constructors

CopyJob'

Fields

iamRoleArn :: Maybe Text

Specifies the IAM role ARN used to copy the target recovery point; for example, arn:aws:iam::123456789012:role/S3Access.

state :: Maybe CopyJobState

The current state of a copy job.

sourceRecoveryPointArn :: Maybe Text

An ARN that uniquely identifies a source recovery point; for example, arn:aws:backup:us-east-1:123456789012:recovery-point:1EB3B5E7-9EB0-435A-A80B-108B488B0D45.

resourceType :: Maybe Text

The type of Amazon Web Services resource to be copied; for example, an Amazon Elastic Block Store (Amazon EBS) volume or an Amazon Relational Database Service (Amazon RDS) database.

destinationBackupVaultArn :: Maybe Text

An Amazon Resource Name (ARN) that uniquely identifies a destination copy vault; for example, arn:aws:backup:us-east-1:123456789012:vault:aBackupVault.

createdBy :: Maybe RecoveryPointCreator

destinationRecoveryPointArn :: Maybe Text

An ARN that uniquely identifies a destination recovery point; for example, arn:aws:backup:us-east-1:123456789012:recovery-point:1EB3B5E7-9EB0-435A-A80B-108B488B0D45.

accountId :: Maybe Text

The account ID that owns the copy job.

sourceBackupVaultArn :: Maybe Text

An Amazon Resource Name (ARN) that uniquely identifies a source copy vault; for example, arn:aws:backup:us-east-1:123456789012:vault:aBackupVault.

copyJobId :: Maybe Text

Uniquely identifies a copy job.

resourceArn :: Maybe Text

The Amazon Web Services resource to be copied; for example, an Amazon Elastic Block Store (Amazon EBS) volume or an Amazon Relational Database Service (Amazon RDS) database.

statusMessage :: Maybe Text

A detailed message explaining the status of the job to copy a resource.

backupSizeInBytes :: Maybe Integer

The size, in bytes, of a copy job.

creationDate :: Maybe POSIX

The date and time a copy job is created, in Unix format and Coordinated Universal Time (UTC). The value of CreationDate is accurate to milliseconds. For example, the value 1516925490.087 represents Friday, January 26, 2018 12:11:30.087 AM.

completionDate :: Maybe POSIX

The date and time a copy job is completed, in Unix format and Coordinated Universal Time (UTC). The value of CompletionDate is accurate to milliseconds. For example, the value 1516925490.087 represents Friday, January 26, 2018 12:11:30.087 AM.

newCopyJob :: CopyJob

Create a value of CopyJob with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:iamRoleArn:CopyJob', copyJob_iamRoleArn - Specifies the IAM role ARN used to copy the target recovery point; for example, arn:aws:iam::123456789012:role/S3Access.

$sel:state:CopyJob', copyJob_state - The current state of a copy job.

$sel:sourceRecoveryPointArn:CopyJob', copyJob_sourceRecoveryPointArn - An ARN that uniquely identifies a source recovery point; for example, arn:aws:backup:us-east-1:123456789012:recovery-point:1EB3B5E7-9EB0-435A-A80B-108B488B0D45.

$sel:resourceType:CopyJob', copyJob_resourceType - The type of Amazon Web Services resource to be copied; for example, an Amazon Elastic Block Store (Amazon EBS) volume or an Amazon Relational Database Service (Amazon RDS) database.

$sel:destinationBackupVaultArn:CopyJob', copyJob_destinationBackupVaultArn - An Amazon Resource Name (ARN) that uniquely identifies a destination copy vault; for example, arn:aws:backup:us-east-1:123456789012:vault:aBackupVault.

$sel:createdBy:CopyJob', copyJob_createdBy - Undocumented member.

$sel:destinationRecoveryPointArn:CopyJob', copyJob_destinationRecoveryPointArn - An ARN that uniquely identifies a destination recovery point; for example, arn:aws:backup:us-east-1:123456789012:recovery-point:1EB3B5E7-9EB0-435A-A80B-108B488B0D45.

$sel:accountId:CopyJob', copyJob_accountId - The account ID that owns the copy job.

$sel:sourceBackupVaultArn:CopyJob', copyJob_sourceBackupVaultArn - An Amazon Resource Name (ARN) that uniquely identifies a source copy vault; for example, arn:aws:backup:us-east-1:123456789012:vault:aBackupVault.

$sel:copyJobId:CopyJob', copyJob_copyJobId - Uniquely identifies a copy job.

$sel:resourceArn:CopyJob', copyJob_resourceArn - The Amazon Web Services resource to be copied; for example, an Amazon Elastic Block Store (Amazon EBS) volume or an Amazon Relational Database Service (Amazon RDS) database.

$sel:statusMessage:CopyJob', copyJob_statusMessage - A detailed message explaining the status of the job to copy a resource.

$sel:backupSizeInBytes:CopyJob', copyJob_backupSizeInBytes - The size, in bytes, of a copy job.

$sel:creationDate:CopyJob', copyJob_creationDate - The date and time a copy job is created, in Unix format and Coordinated Universal Time (UTC). The value of CreationDate is accurate to milliseconds. For example, the value 1516925490.087 represents Friday, January 26, 2018 12:11:30.087 AM.

$sel:completionDate:CopyJob', copyJob_completionDate - The date and time a copy job is completed, in Unix format and Coordinated Universal Time (UTC). The value of CompletionDate is accurate to milliseconds. For example, the value 1516925490.087 represents Friday, January 26, 2018 12:11:30.087 AM.

copyJob_iamRoleArn :: Lens' CopyJob (Maybe Text)

Specifies the IAM role ARN used to copy the target recovery point; for example, arn:aws:iam::123456789012:role/S3Access.

copyJob_state :: Lens' CopyJob (Maybe CopyJobState)

The current state of a copy job.

copyJob_sourceRecoveryPointArn :: Lens' CopyJob (Maybe Text)

An ARN that uniquely identifies a source recovery point; for example, arn:aws:backup:us-east-1:123456789012:recovery-point:1EB3B5E7-9EB0-435A-A80B-108B488B0D45.

copyJob_resourceType :: Lens' CopyJob (Maybe Text)

The type of Amazon Web Services resource to be copied; for example, an Amazon Elastic Block Store (Amazon EBS) volume or an Amazon Relational Database Service (Amazon RDS) database.

copyJob_destinationBackupVaultArn :: Lens' CopyJob (Maybe Text)

An Amazon Resource Name (ARN) that uniquely identifies a destination copy vault; for example, arn:aws:backup:us-east-1:123456789012:vault:aBackupVault.

copyJob_createdBy :: Lens' CopyJob (Maybe RecoveryPointCreator)

Undocumented member.

copyJob_destinationRecoveryPointArn :: Lens' CopyJob (Maybe Text)

An ARN that uniquely identifies a destination recovery point; for example, arn:aws:backup:us-east-1:123456789012:recovery-point:1EB3B5E7-9EB0-435A-A80B-108B488B0D45.

copyJob_accountId :: Lens' CopyJob (Maybe Text)

The account ID that owns the copy job.

copyJob_sourceBackupVaultArn :: Lens' CopyJob (Maybe Text)

An Amazon Resource Name (ARN) that uniquely identifies a source copy vault; for example, arn:aws:backup:us-east-1:123456789012:vault:aBackupVault.

copyJob_copyJobId :: Lens' CopyJob (Maybe Text)

Uniquely identifies a copy job.

copyJob_resourceArn :: Lens' CopyJob (Maybe Text)

The Amazon Web Services resource to be copied; for example, an Amazon Elastic Block Store (Amazon EBS) volume or an Amazon Relational Database Service (Amazon RDS) database.

copyJob_statusMessage :: Lens' CopyJob (Maybe Text)

A detailed message explaining the status of the job to copy a resource.

copyJob_backupSizeInBytes :: Lens' CopyJob (Maybe Integer)

The size, in bytes, of a copy job.

copyJob_creationDate :: Lens' CopyJob (Maybe UTCTime)

The date and time a copy job is created, in Unix format and Coordinated Universal Time (UTC). The value of CreationDate is accurate to milliseconds. For example, the value 1516925490.087 represents Friday, January 26, 2018 12:11:30.087 AM.

copyJob_completionDate :: Lens' CopyJob (Maybe UTCTime)

The date and time a copy job is completed, in Unix format and Coordinated Universal Time (UTC). The value of CompletionDate is accurate to milliseconds. For example, the value 1516925490.087 represents Friday, January 26, 2018 12:11:30.087 AM.