Amazonka.Backup.Types.Lifecycle

Description

Documentation

data Lifecycle

Contains an array of Transition objects specifying how long in days before a recovery point transitions to cold storage or is deleted.

Backups transitioned to cold storage must be stored in cold storage for a minimum of 90 days. Therefore, on the console, the “expire after days” setting must be 90 days greater than the “transition to cold after days” setting. The “transition to cold after days” setting cannot be changed after a backup has been transitioned to cold.

Only Amazon EFS file system backups can be transitioned to cold storage.

See: newLifecycle smart constructor.

Constructors

Lifecycle'

Fields

moveToColdStorageAfterDays :: Maybe Integer

Specifies the number of days after creation that a recovery point is moved to cold storage.

deleteAfterDays :: Maybe Integer

Specifies the number of days after creation that a recovery point is deleted. Must be greater than 90 days plus MoveToColdStorageAfterDays.

newLifecycle :: Lifecycle

Create a value of Lifecycle with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:moveToColdStorageAfterDays:Lifecycle', lifecycle_moveToColdStorageAfterDays - Specifies the number of days after creation that a recovery point is moved to cold storage.

$sel:deleteAfterDays:Lifecycle', lifecycle_deleteAfterDays - Specifies the number of days after creation that a recovery point is deleted. Must be greater than 90 days plus MoveToColdStorageAfterDays.

lifecycle_moveToColdStorageAfterDays :: Lens' Lifecycle (Maybe Integer)

Specifies the number of days after creation that a recovery point is moved to cold storage.

lifecycle_deleteAfterDays :: Lens' Lifecycle (Maybe Integer)

Specifies the number of days after creation that a recovery point is deleted. Must be greater than 90 days plus MoveToColdStorageAfterDays.