Amazonka.Backup.Types.RecoveryPointByResource

Description

Documentation

data RecoveryPointByResource

Contains detailed information about a saved recovery point.

See: newRecoveryPointByResource smart constructor.

Constructors

RecoveryPointByResource'

Fields

status :: Maybe RecoveryPointStatus

A status code specifying the state of the recovery point.

statusMessage :: Maybe Text

A message explaining the reason of the recovery point deletion failure.

recoveryPointArn :: Maybe Text

An Amazon Resource Name (ARN) that uniquely identifies a recovery point; for example, arn:aws:backup:us-east-1:123456789012:recovery-point:1EB3B5E7-9EB0-435A-A80B-108B488B0D45.

backupSizeBytes :: Maybe Integer

The size, in bytes, of a backup.

encryptionKeyArn :: Maybe Text

The server-side encryption key that is used to protect your backups; for example, arn:aws:kms:us-west-2:111122223333:key/1234abcd-12ab-34cd-56ef-1234567890ab.

creationDate :: Maybe POSIX

The date and time a recovery point is created, in Unix format and Coordinated Universal Time (UTC). The value of CreationDate is accurate to milliseconds. For example, the value 1516925490.087 represents Friday, January 26, 2018 12:11:30.087 AM.

backupVaultName :: Maybe Text

The name of a logical container where backups are stored. Backup vaults are identified by names that are unique to the account used to create them and the Amazon Web Services Region where they are created. They consist of lowercase letters, numbers, and hyphens.

newRecoveryPointByResource :: RecoveryPointByResource

Create a value of RecoveryPointByResource with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:status:RecoveryPointByResource', recoveryPointByResource_status - A status code specifying the state of the recovery point.

$sel:statusMessage:RecoveryPointByResource', recoveryPointByResource_statusMessage - A message explaining the reason of the recovery point deletion failure.

$sel:recoveryPointArn:RecoveryPointByResource', recoveryPointByResource_recoveryPointArn - An Amazon Resource Name (ARN) that uniquely identifies a recovery point; for example, arn:aws:backup:us-east-1:123456789012:recovery-point:1EB3B5E7-9EB0-435A-A80B-108B488B0D45.

$sel:backupSizeBytes:RecoveryPointByResource', recoveryPointByResource_backupSizeBytes - The size, in bytes, of a backup.

$sel:encryptionKeyArn:RecoveryPointByResource', recoveryPointByResource_encryptionKeyArn - The server-side encryption key that is used to protect your backups; for example, arn:aws:kms:us-west-2:111122223333:key/1234abcd-12ab-34cd-56ef-1234567890ab.

$sel:creationDate:RecoveryPointByResource', recoveryPointByResource_creationDate - The date and time a recovery point is created, in Unix format and Coordinated Universal Time (UTC). The value of CreationDate is accurate to milliseconds. For example, the value 1516925490.087 represents Friday, January 26, 2018 12:11:30.087 AM.

$sel:backupVaultName:RecoveryPointByResource', recoveryPointByResource_backupVaultName - The name of a logical container where backups are stored. Backup vaults are identified by names that are unique to the account used to create them and the Amazon Web Services Region where they are created. They consist of lowercase letters, numbers, and hyphens.

recoveryPointByResource_status :: Lens' RecoveryPointByResource (Maybe RecoveryPointStatus)

A status code specifying the state of the recovery point.

recoveryPointByResource_statusMessage :: Lens' RecoveryPointByResource (Maybe Text)

A message explaining the reason of the recovery point deletion failure.

recoveryPointByResource_recoveryPointArn :: Lens' RecoveryPointByResource (Maybe Text)

An Amazon Resource Name (ARN) that uniquely identifies a recovery point; for example, arn:aws:backup:us-east-1:123456789012:recovery-point:1EB3B5E7-9EB0-435A-A80B-108B488B0D45.

recoveryPointByResource_backupSizeBytes :: Lens' RecoveryPointByResource (Maybe Integer)

The size, in bytes, of a backup.

recoveryPointByResource_encryptionKeyArn :: Lens' RecoveryPointByResource (Maybe Text)

The server-side encryption key that is used to protect your backups; for example, arn:aws:kms:us-west-2:111122223333:key/1234abcd-12ab-34cd-56ef-1234567890ab.

recoveryPointByResource_creationDate :: Lens' RecoveryPointByResource (Maybe UTCTime)

The date and time a recovery point is created, in Unix format and Coordinated Universal Time (UTC). The value of CreationDate is accurate to milliseconds. For example, the value 1516925490.087 represents Friday, January 26, 2018 12:11:30.087 AM.

recoveryPointByResource_backupVaultName :: Lens' RecoveryPointByResource (Maybe Text)

The name of a logical container where backups are stored. Backup vaults are identified by names that are unique to the account used to create them and the Amazon Web Services Region where they are created. They consist of lowercase letters, numbers, and hyphens.