Amazonka.Backup.Types.ReportJob

Description

Documentation

data ReportJob

Contains detailed information about a report job. A report job compiles a report based on a report plan and publishes it to Amazon S3.

See: newReportJob smart constructor.

Constructors

ReportJob'

Fields

creationTime :: Maybe POSIX

The date and time that a report job is created, in Unix format and Coordinated Universal Time (UTC). The value of CreationTime is accurate to milliseconds. For example, the value 1516925490.087 represents Friday, January 26, 2018 12:11:30.087 AM.

status :: Maybe Text

The status of a report job. The statuses are:

CREATED | RUNNING | COMPLETED | FAILED

COMPLETED means that the report is available for your review at your designated destination. If the status is FAILED, review the StatusMessage for the reason.

reportPlanArn :: Maybe Text

An Amazon Resource Name (ARN) that uniquely identifies a resource. The format of the ARN depends on the resource type.

completionTime :: Maybe POSIX

The date and time that a report job is completed, in Unix format and Coordinated Universal Time (UTC). The value of CompletionTime is accurate to milliseconds. For example, the value 1516925490.087 represents Friday, January 26, 2018 12:11:30.087 AM.

reportJobId :: Maybe Text

The identifier for a report job. A unique, randomly generated, Unicode, UTF-8 encoded string that is at most 1,024 bytes long. Report job IDs cannot be edited.

statusMessage :: Maybe Text

A message explaining the status of the report job.

reportDestination :: Maybe ReportDestination

The S3 bucket name and S3 keys for the destination where the report job publishes the report.

reportTemplate :: Maybe Text

Identifies the report template for the report. Reports are built using a report template. The report templates are:

RESOURCE_COMPLIANCE_REPORT | CONTROL_COMPLIANCE_REPORT | BACKUP_JOB_REPORT | COPY_JOB_REPORT | RESTORE_JOB_REPORT

newReportJob :: ReportJob

Create a value of ReportJob with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:creationTime:ReportJob', reportJob_creationTime - The date and time that a report job is created, in Unix format and Coordinated Universal Time (UTC). The value of CreationTime is accurate to milliseconds. For example, the value 1516925490.087 represents Friday, January 26, 2018 12:11:30.087 AM.

$sel:status:ReportJob', reportJob_status - The status of a report job. The statuses are:

CREATED | RUNNING | COMPLETED | FAILED

COMPLETED means that the report is available for your review at your designated destination. If the status is FAILED, review the StatusMessage for the reason.

$sel:reportPlanArn:ReportJob', reportJob_reportPlanArn - An Amazon Resource Name (ARN) that uniquely identifies a resource. The format of the ARN depends on the resource type.

$sel:completionTime:ReportJob', reportJob_completionTime - The date and time that a report job is completed, in Unix format and Coordinated Universal Time (UTC). The value of CompletionTime is accurate to milliseconds. For example, the value 1516925490.087 represents Friday, January 26, 2018 12:11:30.087 AM.

$sel:reportJobId:ReportJob', reportJob_reportJobId - The identifier for a report job. A unique, randomly generated, Unicode, UTF-8 encoded string that is at most 1,024 bytes long. Report job IDs cannot be edited.

$sel:statusMessage:ReportJob', reportJob_statusMessage - A message explaining the status of the report job.

$sel:reportDestination:ReportJob', reportJob_reportDestination - The S3 bucket name and S3 keys for the destination where the report job publishes the report.

$sel:reportTemplate:ReportJob', reportJob_reportTemplate - Identifies the report template for the report. Reports are built using a report template. The report templates are:

RESOURCE_COMPLIANCE_REPORT | CONTROL_COMPLIANCE_REPORT | BACKUP_JOB_REPORT | COPY_JOB_REPORT | RESTORE_JOB_REPORT

reportJob_creationTime :: Lens' ReportJob (Maybe UTCTime)

The date and time that a report job is created, in Unix format and Coordinated Universal Time (UTC). The value of CreationTime is accurate to milliseconds. For example, the value 1516925490.087 represents Friday, January 26, 2018 12:11:30.087 AM.

reportJob_status :: Lens' ReportJob (Maybe Text)

The status of a report job. The statuses are:

CREATED | RUNNING | COMPLETED | FAILED

COMPLETED means that the report is available for your review at your designated destination. If the status is FAILED, review the StatusMessage for the reason.

reportJob_reportPlanArn :: Lens' ReportJob (Maybe Text)

An Amazon Resource Name (ARN) that uniquely identifies a resource. The format of the ARN depends on the resource type.

reportJob_completionTime :: Lens' ReportJob (Maybe UTCTime)

The date and time that a report job is completed, in Unix format and Coordinated Universal Time (UTC). The value of CompletionTime is accurate to milliseconds. For example, the value 1516925490.087 represents Friday, January 26, 2018 12:11:30.087 AM.

reportJob_reportJobId :: Lens' ReportJob (Maybe Text)

The identifier for a report job. A unique, randomly generated, Unicode, UTF-8 encoded string that is at most 1,024 bytes long. Report job IDs cannot be edited.

reportJob_statusMessage :: Lens' ReportJob (Maybe Text)

A message explaining the status of the report job.

reportJob_reportDestination :: Lens' ReportJob (Maybe ReportDestination)

The S3 bucket name and S3 keys for the destination where the report job publishes the report.

reportJob_reportTemplate :: Lens' ReportJob (Maybe Text)

Identifies the report template for the report. Reports are built using a report template. The report templates are:

RESOURCE_COMPLIANCE_REPORT | CONTROL_COMPLIANCE_REPORT | BACKUP_JOB_REPORT | COPY_JOB_REPORT | RESTORE_JOB_REPORT