Amazonka.Batch.CreateJobQueue

Description

Creates an Batch job queue. When you create a job queue, you associate one or more compute environments to the queue and assign an order of preference for the compute environments.

You also set a priority to the job queue that determines the order that the Batch scheduler places jobs onto its associated compute environments. For example, if a compute environment is associated with more than one job queue, the job queue with a higher priority is given preference for scheduling jobs to that compute environment.

Creating a Request

data CreateJobQueue

Contains the parameters for CreateJobQueue.

See: newCreateJobQueue smart constructor.

Constructors

CreateJobQueue'

Fields

state :: Maybe JQState

The state of the job queue. If the job queue state is ENABLED, it is able to accept jobs. If the job queue state is DISABLED, new jobs can't be added to the queue, but jobs already in the queue can finish.

tags :: Maybe (HashMap Text Text)

The tags that you apply to the job queue to help you categorize and organize your resources. Each tag consists of a key and an optional value. For more information, see Tagging your Batch resources in Batch User Guide.

jobQueueName :: Text

The name of the job queue. Up to 128 letters (uppercase and lowercase), numbers, and underscores are allowed.

priority :: Int

The priority of the job queue. Job queues with a higher priority (or a higher integer value for the priority parameter) are evaluated first when associated with the same compute environment. Priority is determined in descending order. For example, a job queue with a priority value of 10 is given scheduling preference over a job queue with a priority value of 1. All of the compute environments must be either EC2 (EC2 or SPOT) or Fargate (FARGATE or FARGATE_SPOT); EC2 and Fargate compute environments can't be mixed.

computeEnvironmentOrder :: [ComputeEnvironmentOrder]

The set of compute environments mapped to a job queue and their order relative to each other. The job scheduler uses this parameter to determine which compute environment should run a specific job. Compute environments must be in the VALID state before you can associate them with a job queue. You can associate up to three compute environments with a job queue. All of the compute environments must be either EC2 (EC2 or SPOT) or Fargate (FARGATE or FARGATE_SPOT); EC2 and Fargate compute environments can't be mixed.

All compute environments that are associated with a job queue must share the same architecture. Batch doesn't support mixing compute environment architecture types in a single job queue.

newCreateJobQueue

Arguments

:: Text
$sel:jobQueueName:CreateJobQueue'

-> Int
$sel:priority:CreateJobQueue'

-> CreateJobQueue

Create a value of CreateJobQueue with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:state:CreateJobQueue', createJobQueue_state - The state of the job queue. If the job queue state is ENABLED, it is able to accept jobs. If the job queue state is DISABLED, new jobs can't be added to the queue, but jobs already in the queue can finish.

$sel:tags:CreateJobQueue', createJobQueue_tags - The tags that you apply to the job queue to help you categorize and organize your resources. Each tag consists of a key and an optional value. For more information, see Tagging your Batch resources in Batch User Guide.

$sel:jobQueueName:CreateJobQueue', createJobQueue_jobQueueName - The name of the job queue. Up to 128 letters (uppercase and lowercase), numbers, and underscores are allowed.

$sel:priority:CreateJobQueue', createJobQueue_priority - The priority of the job queue. Job queues with a higher priority (or a higher integer value for the priority parameter) are evaluated first when associated with the same compute environment. Priority is determined in descending order. For example, a job queue with a priority value of 10 is given scheduling preference over a job queue with a priority value of 1. All of the compute environments must be either EC2 (EC2 or SPOT) or Fargate (FARGATE or FARGATE_SPOT); EC2 and Fargate compute environments can't be mixed.

$sel:computeEnvironmentOrder:CreateJobQueue', createJobQueue_computeEnvironmentOrder - The set of compute environments mapped to a job queue and their order relative to each other. The job scheduler uses this parameter to determine which compute environment should run a specific job. Compute environments must be in the VALID state before you can associate them with a job queue. You can associate up to three compute environments with a job queue. All of the compute environments must be either EC2 (EC2 or SPOT) or Fargate (FARGATE or FARGATE_SPOT); EC2 and Fargate compute environments can't be mixed.

All compute environments that are associated with a job queue must share the same architecture. Batch doesn't support mixing compute environment architecture types in a single job queue.

Request Lenses

createJobQueue_state :: Lens' CreateJobQueue (Maybe JQState)

The state of the job queue. If the job queue state is ENABLED, it is able to accept jobs. If the job queue state is DISABLED, new jobs can't be added to the queue, but jobs already in the queue can finish.

createJobQueue_tags :: Lens' CreateJobQueue (Maybe (HashMap Text Text))

The tags that you apply to the job queue to help you categorize and organize your resources. Each tag consists of a key and an optional value. For more information, see Tagging your Batch resources in Batch User Guide.

createJobQueue_jobQueueName :: Lens' CreateJobQueue Text

The name of the job queue. Up to 128 letters (uppercase and lowercase), numbers, and underscores are allowed.

createJobQueue_priority :: Lens' CreateJobQueue Int

The priority of the job queue. Job queues with a higher priority (or a higher integer value for the priority parameter) are evaluated first when associated with the same compute environment. Priority is determined in descending order. For example, a job queue with a priority value of 10 is given scheduling preference over a job queue with a priority value of 1. All of the compute environments must be either EC2 (EC2 or SPOT) or Fargate (FARGATE or FARGATE_SPOT); EC2 and Fargate compute environments can't be mixed.

createJobQueue_computeEnvironmentOrder :: Lens' CreateJobQueue [ComputeEnvironmentOrder]

The set of compute environments mapped to a job queue and their order relative to each other. The job scheduler uses this parameter to determine which compute environment should run a specific job. Compute environments must be in the VALID state before you can associate them with a job queue. You can associate up to three compute environments with a job queue. All of the compute environments must be either EC2 (EC2 or SPOT) or Fargate (FARGATE or FARGATE_SPOT); EC2 and Fargate compute environments can't be mixed.

All compute environments that are associated with a job queue must share the same architecture. Batch doesn't support mixing compute environment architecture types in a single job queue.

Destructuring the Response

data CreateJobQueueResponse

See: newCreateJobQueueResponse smart constructor.

Constructors

CreateJobQueueResponse'

Fields

httpStatus :: Int

The response's http status code.

jobQueueName :: Text

The name of the job queue.

jobQueueArn :: Text

The Amazon Resource Name (ARN) of the job queue.

newCreateJobQueueResponse

Arguments

:: Int
$sel:httpStatus:CreateJobQueueResponse'

-> Text
$sel:jobQueueName:CreateJobQueue'

-> Text
$sel:jobQueueArn:CreateJobQueueResponse'

-> CreateJobQueueResponse

Create a value of CreateJobQueueResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:httpStatus:CreateJobQueueResponse', createJobQueueResponse_httpStatus - The response's http status code.

$sel:jobQueueName:CreateJobQueue', createJobQueueResponse_jobQueueName - The name of the job queue.

$sel:jobQueueArn:CreateJobQueueResponse', createJobQueueResponse_jobQueueArn - The Amazon Resource Name (ARN) of the job queue.

Response Lenses

createJobQueueResponse_httpStatus :: Lens' CreateJobQueueResponse Int

The response's http status code.

createJobQueueResponse_jobQueueName :: Lens' CreateJobQueueResponse Text

The name of the job queue.

createJobQueueResponse_jobQueueArn :: Lens' CreateJobQueueResponse Text

The Amazon Resource Name (ARN) of the job queue.