Amazonka.Batch.SubmitJob

Description

Submits an Batch job from a job definition. Parameters that are specified during SubmitJob override parameters defined in the job definition. vCPU and memory requirements that are specified in the ResourceRequirements objects in the job definition are the exception. They can't be overridden this way using the memory and vcpus parameters. Rather, you must specify updates to job definition parameters in a ResourceRequirements object that's included in the containerOverrides parameter.

Jobs that run on Fargate resources can't be guaranteed to run for more than 14 days. This is because, after 14 days, Fargate resources might become unavailable and job might be terminated.

Creating a Request

data SubmitJob

Contains the parameters for SubmitJob.

See: newSubmitJob smart constructor.

Constructors

SubmitJob'

Fields

nodeOverrides :: Maybe NodeOverrides

A list of node overrides in JSON format that specify the node range to target and the container overrides for that node range.

This parameter isn't applicable to jobs that are running on Fargate resources; use containerOverrides instead.

propagateTags :: Maybe Bool

Specifies whether to propagate the tags from the job or job definition to the corresponding Amazon ECS task. If no value is specified, the tags aren't propagated. Tags can only be propagated to the tasks during task creation. For tags with the same name, job tags are given priority over job definitions tags. If the total number of combined tags from the job and job definition is over 50, the job is moved to the FAILED state. When specified, this overrides the tag propagation setting in the job definition.

containerOverrides :: Maybe ContainerOverrides

A list of container overrides in the JSON format that specify the name of a container in the specified job definition and the overrides it should receive. You can override the default command for a container, which is specified in the job definition or the Docker image, with a command override. You can also override existing environment variables on a container or add new environment variables to it with an environment override.

retryStrategy :: Maybe RetryStrategy

The retry strategy to use for failed jobs from this SubmitJob operation. When a retry strategy is specified here, it overrides the retry strategy defined in the job definition.

dependsOn :: Maybe [JobDependency]

A list of dependencies for the job. A job can depend upon a maximum of 20 jobs. You can specify a SEQUENTIAL type dependency without specifying a job ID for array jobs so that each child array job completes sequentially, starting at index 0. You can also specify an N_TO_N type dependency with a job ID for array jobs. In that case, each index child of this job must wait for the corresponding index child of each dependency to complete before it can begin.

parameters :: Maybe (HashMap Text Text)

Additional parameters passed to the job that replace parameter substitution placeholders that are set in the job definition. Parameters are specified as a key and value pair mapping. Parameters in a SubmitJob request override any corresponding parameter defaults from the job definition.

arrayProperties :: Maybe ArrayProperties

The array properties for the submitted job, such as the size of the array. The array size can be between 2 and 10,000. If you specify array properties for a job, it becomes an array job. For more information, see Array Jobs in the Batch User Guide.

timeout :: Maybe JobTimeout

The timeout configuration for this SubmitJob operation. You can specify a timeout duration after which Batch terminates your jobs if they haven't finished. If a job is terminated due to a timeout, it isn't retried. The minimum value for the timeout is 60 seconds. This configuration overrides any timeout configuration specified in the job definition. For array jobs, child jobs have the same timeout configuration as the parent job. For more information, see Job Timeouts in the Amazon Elastic Container Service Developer Guide.

tags :: Maybe (HashMap Text Text)

The tags that you apply to the job request to help you categorize and organize your resources. Each tag consists of a key and an optional value. For more information, see Tagging Amazon Web Services Resources in Amazon Web Services General Reference.

jobName :: Text

The name of the job. The first character must be alphanumeric, and up to 128 letters (uppercase and lowercase), numbers, hyphens, and underscores are allowed.

jobQueue :: Text

The job queue where the job is submitted. You can specify either the name or the Amazon Resource Name (ARN) of the queue.

jobDefinition :: Text

The job definition used by this job. This value can be one of name, name:revision, or the Amazon Resource Name (ARN) for the job definition. If name is specified without a revision then the latest active revision is used.

newSubmitJob

Arguments

:: Text
$sel:jobName:SubmitJob'

-> Text
$sel:jobQueue:SubmitJob'

-> Text
$sel:jobDefinition:SubmitJob'

-> SubmitJob

Create a value of SubmitJob with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:nodeOverrides:SubmitJob', submitJob_nodeOverrides - A list of node overrides in JSON format that specify the node range to target and the container overrides for that node range.

This parameter isn't applicable to jobs that are running on Fargate resources; use containerOverrides instead.

$sel:propagateTags:SubmitJob', submitJob_propagateTags - Specifies whether to propagate the tags from the job or job definition to the corresponding Amazon ECS task. If no value is specified, the tags aren't propagated. Tags can only be propagated to the tasks during task creation. For tags with the same name, job tags are given priority over job definitions tags. If the total number of combined tags from the job and job definition is over 50, the job is moved to the FAILED state. When specified, this overrides the tag propagation setting in the job definition.

$sel:containerOverrides:SubmitJob', submitJob_containerOverrides - A list of container overrides in the JSON format that specify the name of a container in the specified job definition and the overrides it should receive. You can override the default command for a container, which is specified in the job definition or the Docker image, with a command override. You can also override existing environment variables on a container or add new environment variables to it with an environment override.

$sel:retryStrategy:SubmitJob', submitJob_retryStrategy - The retry strategy to use for failed jobs from this SubmitJob operation. When a retry strategy is specified here, it overrides the retry strategy defined in the job definition.

$sel:dependsOn:SubmitJob', submitJob_dependsOn - A list of dependencies for the job. A job can depend upon a maximum of 20 jobs. You can specify a SEQUENTIAL type dependency without specifying a job ID for array jobs so that each child array job completes sequentially, starting at index 0. You can also specify an N_TO_N type dependency with a job ID for array jobs. In that case, each index child of this job must wait for the corresponding index child of each dependency to complete before it can begin.

$sel:parameters:SubmitJob', submitJob_parameters - Additional parameters passed to the job that replace parameter substitution placeholders that are set in the job definition. Parameters are specified as a key and value pair mapping. Parameters in a SubmitJob request override any corresponding parameter defaults from the job definition.

$sel:arrayProperties:SubmitJob', submitJob_arrayProperties - The array properties for the submitted job, such as the size of the array. The array size can be between 2 and 10,000. If you specify array properties for a job, it becomes an array job. For more information, see Array Jobs in the Batch User Guide.

$sel:timeout:SubmitJob', submitJob_timeout - The timeout configuration for this SubmitJob operation. You can specify a timeout duration after which Batch terminates your jobs if they haven't finished. If a job is terminated due to a timeout, it isn't retried. The minimum value for the timeout is 60 seconds. This configuration overrides any timeout configuration specified in the job definition. For array jobs, child jobs have the same timeout configuration as the parent job. For more information, see Job Timeouts in the Amazon Elastic Container Service Developer Guide.

$sel:tags:SubmitJob', submitJob_tags - The tags that you apply to the job request to help you categorize and organize your resources. Each tag consists of a key and an optional value. For more information, see Tagging Amazon Web Services Resources in Amazon Web Services General Reference.

$sel:jobName:SubmitJob', submitJob_jobName - The name of the job. The first character must be alphanumeric, and up to 128 letters (uppercase and lowercase), numbers, hyphens, and underscores are allowed.

$sel:jobQueue:SubmitJob', submitJob_jobQueue - The job queue where the job is submitted. You can specify either the name or the Amazon Resource Name (ARN) of the queue.

$sel:jobDefinition:SubmitJob', submitJob_jobDefinition - The job definition used by this job. This value can be one of name, name:revision, or the Amazon Resource Name (ARN) for the job definition. If name is specified without a revision then the latest active revision is used.

Request Lenses

submitJob_nodeOverrides :: Lens' SubmitJob (Maybe NodeOverrides)

A list of node overrides in JSON format that specify the node range to target and the container overrides for that node range.

This parameter isn't applicable to jobs that are running on Fargate resources; use containerOverrides instead.

submitJob_propagateTags :: Lens' SubmitJob (Maybe Bool)

Specifies whether to propagate the tags from the job or job definition to the corresponding Amazon ECS task. If no value is specified, the tags aren't propagated. Tags can only be propagated to the tasks during task creation. For tags with the same name, job tags are given priority over job definitions tags. If the total number of combined tags from the job and job definition is over 50, the job is moved to the FAILED state. When specified, this overrides the tag propagation setting in the job definition.

submitJob_containerOverrides :: Lens' SubmitJob (Maybe ContainerOverrides)

A list of container overrides in the JSON format that specify the name of a container in the specified job definition and the overrides it should receive. You can override the default command for a container, which is specified in the job definition or the Docker image, with a command override. You can also override existing environment variables on a container or add new environment variables to it with an environment override.

submitJob_retryStrategy :: Lens' SubmitJob (Maybe RetryStrategy)

The retry strategy to use for failed jobs from this SubmitJob operation. When a retry strategy is specified here, it overrides the retry strategy defined in the job definition.

submitJob_dependsOn :: Lens' SubmitJob (Maybe [JobDependency])

A list of dependencies for the job. A job can depend upon a maximum of 20 jobs. You can specify a SEQUENTIAL type dependency without specifying a job ID for array jobs so that each child array job completes sequentially, starting at index 0. You can also specify an N_TO_N type dependency with a job ID for array jobs. In that case, each index child of this job must wait for the corresponding index child of each dependency to complete before it can begin.

submitJob_parameters :: Lens' SubmitJob (Maybe (HashMap Text Text))

Additional parameters passed to the job that replace parameter substitution placeholders that are set in the job definition. Parameters are specified as a key and value pair mapping. Parameters in a SubmitJob request override any corresponding parameter defaults from the job definition.

submitJob_arrayProperties :: Lens' SubmitJob (Maybe ArrayProperties)

The array properties for the submitted job, such as the size of the array. The array size can be between 2 and 10,000. If you specify array properties for a job, it becomes an array job. For more information, see Array Jobs in the Batch User Guide.

submitJob_timeout :: Lens' SubmitJob (Maybe JobTimeout)

The timeout configuration for this SubmitJob operation. You can specify a timeout duration after which Batch terminates your jobs if they haven't finished. If a job is terminated due to a timeout, it isn't retried. The minimum value for the timeout is 60 seconds. This configuration overrides any timeout configuration specified in the job definition. For array jobs, child jobs have the same timeout configuration as the parent job. For more information, see Job Timeouts in the Amazon Elastic Container Service Developer Guide.

submitJob_tags :: Lens' SubmitJob (Maybe (HashMap Text Text))

The tags that you apply to the job request to help you categorize and organize your resources. Each tag consists of a key and an optional value. For more information, see Tagging Amazon Web Services Resources in Amazon Web Services General Reference.

submitJob_jobName :: Lens' SubmitJob Text

The name of the job. The first character must be alphanumeric, and up to 128 letters (uppercase and lowercase), numbers, hyphens, and underscores are allowed.

submitJob_jobQueue :: Lens' SubmitJob Text

The job queue where the job is submitted. You can specify either the name or the Amazon Resource Name (ARN) of the queue.

submitJob_jobDefinition :: Lens' SubmitJob Text

The job definition used by this job. This value can be one of name, name:revision, or the Amazon Resource Name (ARN) for the job definition. If name is specified without a revision then the latest active revision is used.

Destructuring the Response

data SubmitJobResponse

See: newSubmitJobResponse smart constructor.

Constructors

SubmitJobResponse'

Fields

jobArn :: Maybe Text

The Amazon Resource Name (ARN) for the job.

httpStatus :: Int

The response's http status code.

jobName :: Text

The name of the job.

jobId :: Text

The unique identifier for the job.

newSubmitJobResponse

Arguments

:: Int
$sel:httpStatus:SubmitJobResponse'

-> Text
$sel:jobName:SubmitJob'

-> Text
$sel:jobId:SubmitJobResponse'

-> SubmitJobResponse

Create a value of SubmitJobResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:jobArn:SubmitJobResponse', submitJobResponse_jobArn - The Amazon Resource Name (ARN) for the job.

$sel:httpStatus:SubmitJobResponse', submitJobResponse_httpStatus - The response's http status code.

$sel:jobName:SubmitJob', submitJobResponse_jobName - The name of the job.

$sel:jobId:SubmitJobResponse', submitJobResponse_jobId - The unique identifier for the job.

Response Lenses

submitJobResponse_jobArn :: Lens' SubmitJobResponse (Maybe Text)

The Amazon Resource Name (ARN) for the job.

submitJobResponse_httpStatus :: Lens' SubmitJobResponse Int

The response's http status code.

submitJobResponse_jobName :: Lens' SubmitJobResponse Text

The name of the job.

submitJobResponse_jobId :: Lens' SubmitJobResponse Text

The unique identifier for the job.