Amazonka.Batch.Types.ContainerDetail

Description

Documentation

data ContainerDetail

An object representing the details of a container that's part of a job.

See: newContainerDetail smart constructor.

Constructors

ContainerDetail'

Fields

image :: Maybe Text

The image used to start the container.

command :: Maybe [Text]

The command that's passed to the container.

secrets :: Maybe [Secret]

The secrets to pass to the container. For more information, see Specifying sensitive data in the Batch User Guide.

environment :: Maybe [KeyValuePair]

The environment variables to pass to a container.

Environment variables must not start with AWS_BATCH; this naming convention is reserved for variables that are set by the Batch service.

networkInterfaces :: Maybe [NetworkInterface]

The network interfaces associated with the job.

taskArn :: Maybe Text

The Amazon Resource Name (ARN) of the Amazon ECS task that's associated with the container job. Each container attempt receives a task ARN when they reach the STARTING status.

ulimits :: Maybe [Ulimit]

A list of ulimit values to set in the container. This parameter maps to Ulimits in the Create a container section of the Docker Remote API and the --ulimit option to docker run.

This parameter isn't applicable to jobs that are running on Fargate resources.

containerInstanceArn :: Maybe Text

The Amazon Resource Name (ARN) of the container instance that the container is running on.

executionRoleArn :: Maybe Text

The Amazon Resource Name (ARN) of the execution role that Batch can assume. For more information, see Batch execution IAM role in the Batch User Guide.

privileged :: Maybe Bool

When this parameter is true, the container is given elevated permissions on the host container instance (similar to the root user). The default value is false.

This parameter isn't applicable to jobs that are running on Fargate resources and shouldn't be provided, or specified as false.

jobRoleArn :: Maybe Text

The Amazon Resource Name (ARN) associated with the job upon execution.

resourceRequirements :: Maybe [ResourceRequirement]

The type and amount of resources to assign to a container. The supported resources include GPU, MEMORY, and VCPU.

instanceType :: Maybe Text

The instance type of the underlying host infrastructure of a multi-node parallel job.

This parameter isn't applicable to jobs that are running on Fargate resources.

memory :: Maybe Int

For jobs run on EC2 resources that didn't specify memory requirements using ResourceRequirement, the number of MiB of memory reserved for the job. For other jobs, including all run on Fargate resources, see resourceRequirements.

user :: Maybe Text

The user name to use inside the container. This parameter maps to User in the Create a container section of the Docker Remote API and the --user option to docker run.

logConfiguration :: Maybe LogConfiguration

The log configuration specification for the container.

This parameter maps to LogConfig in the Create a container section of the Docker Remote API and the --log-driver option to docker run. By default, containers use the same logging driver that the Docker daemon uses. However, the container might use a different logging driver than the Docker daemon by specifying a log driver with this parameter in the container definition. To use a different logging driver for a container, the log system must be configured properly on the container instance. Or, alternatively, it must be configured on a different log server for remote logging options. For more information on the options for different supported log drivers, see Configure logging drivers in the Docker documentation.

Batch currently supports a subset of the logging drivers available to the Docker daemon (shown in the LogConfiguration data type). Additional log drivers might be available in future releases of the Amazon ECS container agent.

This parameter requires version 1.18 of the Docker Remote API or greater on your container instance. To check the Docker Remote API version on your container instance, log into your container instance and run the following command: sudo docker version | grep "Server API version"

The Amazon ECS container agent running on a container instance must register the logging drivers available on that instance with the ECS_AVAILABLE_LOGGING_DRIVERS environment variable before containers placed on that instance can use these log configuration options. For more information, see Amazon ECS Container Agent Configuration in the Amazon Elastic Container Service Developer Guide.

linuxParameters :: Maybe LinuxParameters

Linux-specific modifications that are applied to the container, such as details for device mappings.

reason :: Maybe Text

A short (255 max characters) human-readable string to provide additional details about a running or stopped container.

logStreamName :: Maybe Text

The name of the CloudWatch Logs log stream associated with the container. The log group for Batch jobs is /aws/batch/job. Each container attempt receives a log stream name when they reach the RUNNING status.

mountPoints :: Maybe [MountPoint]

The mount points for data volumes in your container.

exitCode :: Maybe Int

The exit code to return upon completion.

fargatePlatformConfiguration :: Maybe FargatePlatformConfiguration

The platform configuration for jobs that are running on Fargate resources. Jobs that are running on EC2 resources must not specify this parameter.

vcpus :: Maybe Int

The number of vCPUs reserved for the container. For jobs that run on EC2 resources, you can specify the vCPU requirement for the job using resourceRequirements, but you can't specify the vCPU requirements in both the vcpus and resourceRequirement object. This parameter maps to CpuShares in the Create a container section of the Docker Remote API and the --cpu-shares option to docker run. Each vCPU is equivalent to 1,024 CPU shares. You must specify at least one vCPU. This is required but can be specified in several places. It must be specified for each node at least once.

This parameter isn't applicable to jobs that run on Fargate resources. For jobs that run on Fargate resources, you must specify the vCPU requirement for the job using resourceRequirements.

readonlyRootFilesystem :: Maybe Bool

When this parameter is true, the container is given read-only access to its root file system. This parameter maps to ReadonlyRootfs in the Create a container section of the Docker Remote API and the --read-only option to docker run .

volumes :: Maybe [Volume]

A list of volumes associated with the job.

networkConfiguration :: Maybe NetworkConfiguration

The network configuration for jobs that are running on Fargate resources. Jobs that are running on EC2 resources must not specify this parameter.

newContainerDetail :: ContainerDetail

Create a value of ContainerDetail with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:image:ContainerDetail', containerDetail_image - The image used to start the container.

$sel:command:ContainerDetail', containerDetail_command - The command that's passed to the container.

$sel:secrets:ContainerDetail', containerDetail_secrets - The secrets to pass to the container. For more information, see Specifying sensitive data in the Batch User Guide.

$sel:environment:ContainerDetail', containerDetail_environment - The environment variables to pass to a container.

Environment variables must not start with AWS_BATCH; this naming convention is reserved for variables that are set by the Batch service.

$sel:networkInterfaces:ContainerDetail', containerDetail_networkInterfaces - The network interfaces associated with the job.

$sel:taskArn:ContainerDetail', containerDetail_taskArn - The Amazon Resource Name (ARN) of the Amazon ECS task that's associated with the container job. Each container attempt receives a task ARN when they reach the STARTING status.

$sel:ulimits:ContainerDetail', containerDetail_ulimits - A list of ulimit values to set in the container. This parameter maps to Ulimits in the Create a container section of the Docker Remote API and the --ulimit option to docker run.

This parameter isn't applicable to jobs that are running on Fargate resources.

$sel:containerInstanceArn:ContainerDetail', containerDetail_containerInstanceArn - The Amazon Resource Name (ARN) of the container instance that the container is running on.

$sel:executionRoleArn:ContainerDetail', containerDetail_executionRoleArn - The Amazon Resource Name (ARN) of the execution role that Batch can assume. For more information, see Batch execution IAM role in the Batch User Guide.

$sel:privileged:ContainerDetail', containerDetail_privileged - When this parameter is true, the container is given elevated permissions on the host container instance (similar to the root user). The default value is false.

This parameter isn't applicable to jobs that are running on Fargate resources and shouldn't be provided, or specified as false.

$sel:jobRoleArn:ContainerDetail', containerDetail_jobRoleArn - The Amazon Resource Name (ARN) associated with the job upon execution.

$sel:resourceRequirements:ContainerDetail', containerDetail_resourceRequirements - The type and amount of resources to assign to a container. The supported resources include GPU, MEMORY, and VCPU.

$sel:instanceType:ContainerDetail', containerDetail_instanceType - The instance type of the underlying host infrastructure of a multi-node parallel job.

This parameter isn't applicable to jobs that are running on Fargate resources.

$sel:memory:ContainerDetail', containerDetail_memory - For jobs run on EC2 resources that didn't specify memory requirements using ResourceRequirement, the number of MiB of memory reserved for the job. For other jobs, including all run on Fargate resources, see resourceRequirements.

$sel:user:ContainerDetail', containerDetail_user - The user name to use inside the container. This parameter maps to User in the Create a container section of the Docker Remote API and the --user option to docker run.

$sel:logConfiguration:ContainerDetail', containerDetail_logConfiguration - The log configuration specification for the container.

This parameter maps to LogConfig in the Create a container section of the Docker Remote API and the --log-driver option to docker run. By default, containers use the same logging driver that the Docker daemon uses. However, the container might use a different logging driver than the Docker daemon by specifying a log driver with this parameter in the container definition. To use a different logging driver for a container, the log system must be configured properly on the container instance. Or, alternatively, it must be configured on a different log server for remote logging options. For more information on the options for different supported log drivers, see Configure logging drivers in the Docker documentation.

Batch currently supports a subset of the logging drivers available to the Docker daemon (shown in the LogConfiguration data type). Additional log drivers might be available in future releases of the Amazon ECS container agent.

This parameter requires version 1.18 of the Docker Remote API or greater on your container instance. To check the Docker Remote API version on your container instance, log into your container instance and run the following command: sudo docker version | grep "Server API version"

The Amazon ECS container agent running on a container instance must register the logging drivers available on that instance with the ECS_AVAILABLE_LOGGING_DRIVERS environment variable before containers placed on that instance can use these log configuration options. For more information, see Amazon ECS Container Agent Configuration in the Amazon Elastic Container Service Developer Guide.

$sel:linuxParameters:ContainerDetail', containerDetail_linuxParameters - Linux-specific modifications that are applied to the container, such as details for device mappings.

$sel:reason:ContainerDetail', containerDetail_reason - A short (255 max characters) human-readable string to provide additional details about a running or stopped container.

$sel:logStreamName:ContainerDetail', containerDetail_logStreamName - The name of the CloudWatch Logs log stream associated with the container. The log group for Batch jobs is /aws/batch/job. Each container attempt receives a log stream name when they reach the RUNNING status.

$sel:mountPoints:ContainerDetail', containerDetail_mountPoints - The mount points for data volumes in your container.

$sel:exitCode:ContainerDetail', containerDetail_exitCode - The exit code to return upon completion.

$sel:fargatePlatformConfiguration:ContainerDetail', containerDetail_fargatePlatformConfiguration - The platform configuration for jobs that are running on Fargate resources. Jobs that are running on EC2 resources must not specify this parameter.

$sel:vcpus:ContainerDetail', containerDetail_vcpus - The number of vCPUs reserved for the container. For jobs that run on EC2 resources, you can specify the vCPU requirement for the job using resourceRequirements, but you can't specify the vCPU requirements in both the vcpus and resourceRequirement object. This parameter maps to CpuShares in the Create a container section of the Docker Remote API and the --cpu-shares option to docker run. Each vCPU is equivalent to 1,024 CPU shares. You must specify at least one vCPU. This is required but can be specified in several places. It must be specified for each node at least once.

This parameter isn't applicable to jobs that run on Fargate resources. For jobs that run on Fargate resources, you must specify the vCPU requirement for the job using resourceRequirements.

$sel:readonlyRootFilesystem:ContainerDetail', containerDetail_readonlyRootFilesystem - When this parameter is true, the container is given read-only access to its root file system. This parameter maps to ReadonlyRootfs in the Create a container section of the Docker Remote API and the --read-only option to docker run .

$sel:volumes:ContainerDetail', containerDetail_volumes - A list of volumes associated with the job.

$sel:networkConfiguration:ContainerDetail', containerDetail_networkConfiguration - The network configuration for jobs that are running on Fargate resources. Jobs that are running on EC2 resources must not specify this parameter.

containerDetail_image :: Lens' ContainerDetail (Maybe Text)

The image used to start the container.

containerDetail_command :: Lens' ContainerDetail (Maybe [Text])

The command that's passed to the container.

containerDetail_secrets :: Lens' ContainerDetail (Maybe [Secret])

The secrets to pass to the container. For more information, see Specifying sensitive data in the Batch User Guide.

containerDetail_environment :: Lens' ContainerDetail (Maybe [KeyValuePair])

The environment variables to pass to a container.

Environment variables must not start with AWS_BATCH; this naming convention is reserved for variables that are set by the Batch service.

containerDetail_networkInterfaces :: Lens' ContainerDetail (Maybe [NetworkInterface])

The network interfaces associated with the job.

containerDetail_taskArn :: Lens' ContainerDetail (Maybe Text)

The Amazon Resource Name (ARN) of the Amazon ECS task that's associated with the container job. Each container attempt receives a task ARN when they reach the STARTING status.

containerDetail_ulimits :: Lens' ContainerDetail (Maybe [Ulimit])

A list of ulimit values to set in the container. This parameter maps to Ulimits in the Create a container section of the Docker Remote API and the --ulimit option to docker run.

This parameter isn't applicable to jobs that are running on Fargate resources.

containerDetail_containerInstanceArn :: Lens' ContainerDetail (Maybe Text)

The Amazon Resource Name (ARN) of the container instance that the container is running on.

containerDetail_executionRoleArn :: Lens' ContainerDetail (Maybe Text)

The Amazon Resource Name (ARN) of the execution role that Batch can assume. For more information, see Batch execution IAM role in the Batch User Guide.

containerDetail_privileged :: Lens' ContainerDetail (Maybe Bool)

When this parameter is true, the container is given elevated permissions on the host container instance (similar to the root user). The default value is false.

This parameter isn't applicable to jobs that are running on Fargate resources and shouldn't be provided, or specified as false.

containerDetail_jobRoleArn :: Lens' ContainerDetail (Maybe Text)

The Amazon Resource Name (ARN) associated with the job upon execution.

containerDetail_resourceRequirements :: Lens' ContainerDetail (Maybe [ResourceRequirement])

The type and amount of resources to assign to a container. The supported resources include GPU, MEMORY, and VCPU.

containerDetail_instanceType :: Lens' ContainerDetail (Maybe Text)

The instance type of the underlying host infrastructure of a multi-node parallel job.

This parameter isn't applicable to jobs that are running on Fargate resources.

containerDetail_memory :: Lens' ContainerDetail (Maybe Int)

For jobs run on EC2 resources that didn't specify memory requirements using ResourceRequirement, the number of MiB of memory reserved for the job. For other jobs, including all run on Fargate resources, see resourceRequirements.

containerDetail_user :: Lens' ContainerDetail (Maybe Text)

The user name to use inside the container. This parameter maps to User in the Create a container section of the Docker Remote API and the --user option to docker run.

containerDetail_logConfiguration :: Lens' ContainerDetail (Maybe LogConfiguration)

The log configuration specification for the container.

This parameter maps to LogConfig in the Create a container section of the Docker Remote API and the --log-driver option to docker run. By default, containers use the same logging driver that the Docker daemon uses. However, the container might use a different logging driver than the Docker daemon by specifying a log driver with this parameter in the container definition. To use a different logging driver for a container, the log system must be configured properly on the container instance. Or, alternatively, it must be configured on a different log server for remote logging options. For more information on the options for different supported log drivers, see Configure logging drivers in the Docker documentation.

Batch currently supports a subset of the logging drivers available to the Docker daemon (shown in the LogConfiguration data type). Additional log drivers might be available in future releases of the Amazon ECS container agent.

This parameter requires version 1.18 of the Docker Remote API or greater on your container instance. To check the Docker Remote API version on your container instance, log into your container instance and run the following command: sudo docker version | grep "Server API version"

The Amazon ECS container agent running on a container instance must register the logging drivers available on that instance with the ECS_AVAILABLE_LOGGING_DRIVERS environment variable before containers placed on that instance can use these log configuration options. For more information, see Amazon ECS Container Agent Configuration in the Amazon Elastic Container Service Developer Guide.

containerDetail_linuxParameters :: Lens' ContainerDetail (Maybe LinuxParameters)

Linux-specific modifications that are applied to the container, such as details for device mappings.

containerDetail_reason :: Lens' ContainerDetail (Maybe Text)

A short (255 max characters) human-readable string to provide additional details about a running or stopped container.

containerDetail_logStreamName :: Lens' ContainerDetail (Maybe Text)

The name of the CloudWatch Logs log stream associated with the container. The log group for Batch jobs is /aws/batch/job. Each container attempt receives a log stream name when they reach the RUNNING status.

containerDetail_mountPoints :: Lens' ContainerDetail (Maybe [MountPoint])

The mount points for data volumes in your container.

containerDetail_exitCode :: Lens' ContainerDetail (Maybe Int)

The exit code to return upon completion.

containerDetail_fargatePlatformConfiguration :: Lens' ContainerDetail (Maybe FargatePlatformConfiguration)

The platform configuration for jobs that are running on Fargate resources. Jobs that are running on EC2 resources must not specify this parameter.

containerDetail_vcpus :: Lens' ContainerDetail (Maybe Int)

The number of vCPUs reserved for the container. For jobs that run on EC2 resources, you can specify the vCPU requirement for the job using resourceRequirements, but you can't specify the vCPU requirements in both the vcpus and resourceRequirement object. This parameter maps to CpuShares in the Create a container section of the Docker Remote API and the --cpu-shares option to docker run. Each vCPU is equivalent to 1,024 CPU shares. You must specify at least one vCPU. This is required but can be specified in several places. It must be specified for each node at least once.

This parameter isn't applicable to jobs that run on Fargate resources. For jobs that run on Fargate resources, you must specify the vCPU requirement for the job using resourceRequirements.

containerDetail_readonlyRootFilesystem :: Lens' ContainerDetail (Maybe Bool)

When this parameter is true, the container is given read-only access to its root file system. This parameter maps to ReadonlyRootfs in the Create a container section of the Docker Remote API and the --read-only option to docker run .

containerDetail_volumes :: Lens' ContainerDetail (Maybe [Volume])

A list of volumes associated with the job.

containerDetail_networkConfiguration :: Lens' ContainerDetail (Maybe NetworkConfiguration)

The network configuration for jobs that are running on Fargate resources. Jobs that are running on EC2 resources must not specify this parameter.