Amazonka.Batch.Types.NodePropertiesSummary

Description

Documentation

data NodePropertiesSummary

An object representing the properties of a node that's associated with a multi-node parallel job.

See: newNodePropertiesSummary smart constructor.

Constructors

NodePropertiesSummary'

Fields

numNodes :: Maybe Int

The number of nodes associated with a multi-node parallel job.

nodeIndex :: Maybe Int

The node index for the node. Node index numbering begins at zero. This index is also available on the node with the AWS_BATCH_JOB_NODE_INDEX environment variable.

isMainNode :: Maybe Bool

Specifies whether the current node is the main node for a multi-node parallel job.

newNodePropertiesSummary :: NodePropertiesSummary

Create a value of NodePropertiesSummary with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:numNodes:NodePropertiesSummary', nodePropertiesSummary_numNodes - The number of nodes associated with a multi-node parallel job.

$sel:nodeIndex:NodePropertiesSummary', nodePropertiesSummary_nodeIndex - The node index for the node. Node index numbering begins at zero. This index is also available on the node with the AWS_BATCH_JOB_NODE_INDEX environment variable.

$sel:isMainNode:NodePropertiesSummary', nodePropertiesSummary_isMainNode - Specifies whether the current node is the main node for a multi-node parallel job.

nodePropertiesSummary_numNodes :: Lens' NodePropertiesSummary (Maybe Int)

The number of nodes associated with a multi-node parallel job.

nodePropertiesSummary_nodeIndex :: Lens' NodePropertiesSummary (Maybe Int)

The node index for the node. Node index numbering begins at zero. This index is also available on the node with the AWS_BATCH_JOB_NODE_INDEX environment variable.

nodePropertiesSummary_isMainNode :: Lens' NodePropertiesSummary (Maybe Bool)

Specifies whether the current node is the main node for a multi-node parallel job.