Amazonka.Batch.Types.RetryStrategy

Description

Documentation

data RetryStrategy

The retry strategy associated with a job. For more information, see Automated job retries in the Batch User Guide.

See: newRetryStrategy smart constructor.

Constructors

RetryStrategy'

Fields

evaluateOnExit :: Maybe [EvaluateOnExit]

Array of up to 5 objects that specify conditions under which the job should be retried or failed. If this parameter is specified, then the attempts parameter must also be specified.

attempts :: Maybe Int

The number of times to move a job to the RUNNABLE status. You can specify between 1 and 10 attempts. If the value of attempts is greater than one, the job is retried on failure the same number of attempts as the value.

newRetryStrategy :: RetryStrategy

Create a value of RetryStrategy with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:evaluateOnExit:RetryStrategy', retryStrategy_evaluateOnExit - Array of up to 5 objects that specify conditions under which the job should be retried or failed. If this parameter is specified, then the attempts parameter must also be specified.

$sel:attempts:RetryStrategy', retryStrategy_attempts - The number of times to move a job to the RUNNABLE status. You can specify between 1 and 10 attempts. If the value of attempts is greater than one, the job is retried on failure the same number of attempts as the value.

retryStrategy_evaluateOnExit :: Lens' RetryStrategy (Maybe [EvaluateOnExit])

Array of up to 5 objects that specify conditions under which the job should be retried or failed. If this parameter is specified, then the attempts parameter must also be specified.

retryStrategy_attempts :: Lens' RetryStrategy (Maybe Int)

The number of times to move a job to the RUNNABLE status. You can specify between 1 and 10 attempts. If the value of attempts is greater than one, the job is retried on failure the same number of attempts as the value.