Amazonka.Batch.UpdateComputeEnvironment

Description

Updates an Batch compute environment.

Creating a Request

data UpdateComputeEnvironment

Contains the parameters for UpdateComputeEnvironment.

See: newUpdateComputeEnvironment smart constructor.

Constructors

UpdateComputeEnvironment'

Fields

state :: Maybe CEState

The state of the compute environment. Compute environments in the ENABLED state can accept jobs from a queue and scale in or out automatically based on the workload demand of its associated queues.

If the state is ENABLED, then the Batch scheduler can attempt to place jobs from an associated job queue on the compute resources within the environment. If the compute environment is managed, then it can scale its instances out or in automatically, based on the job queue demand.

If the state is DISABLED, then the Batch scheduler doesn't attempt to place jobs within the environment. Jobs in a STARTING or RUNNING state continue to progress normally. Managed compute environments in the DISABLED state don't scale out. However, they scale in to minvCpus value after instances become idle.

computeResources :: Maybe ComputeResourceUpdate

Details of the compute resources managed by the compute environment. Required for a managed compute environment. For more information, see Compute Environments in the Batch User Guide.

serviceRole :: Maybe Text

The full Amazon Resource Name (ARN) of the IAM role that allows Batch to make calls to other Amazon Web Services services on your behalf. For more information, see Batch service IAM role in the Batch User Guide.

If the compute environment has a service-linked role, it can't be changed to use a regular IAM role. Likewise, if the compute environment has a regular IAM role, it can't be changed to use a service-linked role.

If your specified role has a path other than /, then you must either specify the full role ARN (this is recommended) or prefix the role name with the path.

Depending on how you created your Batch service role, its ARN might contain the service-role path prefix. When you only specify the name of the service role, Batch assumes that your ARN doesn't use the service-role path prefix. Because of this, we recommend that you specify the full ARN of your service role when you create compute environments.

computeEnvironment :: Text

The name or full Amazon Resource Name (ARN) of the compute environment to update.

newUpdateComputeEnvironment

Arguments

:: Text
$sel:computeEnvironment:UpdateComputeEnvironment'

-> UpdateComputeEnvironment

Create a value of UpdateComputeEnvironment with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:state:UpdateComputeEnvironment', updateComputeEnvironment_state - The state of the compute environment. Compute environments in the ENABLED state can accept jobs from a queue and scale in or out automatically based on the workload demand of its associated queues.

If the state is ENABLED, then the Batch scheduler can attempt to place jobs from an associated job queue on the compute resources within the environment. If the compute environment is managed, then it can scale its instances out or in automatically, based on the job queue demand.

If the state is DISABLED, then the Batch scheduler doesn't attempt to place jobs within the environment. Jobs in a STARTING or RUNNING state continue to progress normally. Managed compute environments in the DISABLED state don't scale out. However, they scale in to minvCpus value after instances become idle.

$sel:computeResources:UpdateComputeEnvironment', updateComputeEnvironment_computeResources - Details of the compute resources managed by the compute environment. Required for a managed compute environment. For more information, see Compute Environments in the Batch User Guide.

$sel:serviceRole:UpdateComputeEnvironment', updateComputeEnvironment_serviceRole - The full Amazon Resource Name (ARN) of the IAM role that allows Batch to make calls to other Amazon Web Services services on your behalf. For more information, see Batch service IAM role in the Batch User Guide.

If the compute environment has a service-linked role, it can't be changed to use a regular IAM role. Likewise, if the compute environment has a regular IAM role, it can't be changed to use a service-linked role.

If your specified role has a path other than /, then you must either specify the full role ARN (this is recommended) or prefix the role name with the path.

Depending on how you created your Batch service role, its ARN might contain the service-role path prefix. When you only specify the name of the service role, Batch assumes that your ARN doesn't use the service-role path prefix. Because of this, we recommend that you specify the full ARN of your service role when you create compute environments.

$sel:computeEnvironment:UpdateComputeEnvironment', updateComputeEnvironment_computeEnvironment - The name or full Amazon Resource Name (ARN) of the compute environment to update.

Request Lenses

updateComputeEnvironment_state :: Lens' UpdateComputeEnvironment (Maybe CEState)

The state of the compute environment. Compute environments in the ENABLED state can accept jobs from a queue and scale in or out automatically based on the workload demand of its associated queues.

If the state is ENABLED, then the Batch scheduler can attempt to place jobs from an associated job queue on the compute resources within the environment. If the compute environment is managed, then it can scale its instances out or in automatically, based on the job queue demand.

If the state is DISABLED, then the Batch scheduler doesn't attempt to place jobs within the environment. Jobs in a STARTING or RUNNING state continue to progress normally. Managed compute environments in the DISABLED state don't scale out. However, they scale in to minvCpus value after instances become idle.

updateComputeEnvironment_computeResources :: Lens' UpdateComputeEnvironment (Maybe ComputeResourceUpdate)

Details of the compute resources managed by the compute environment. Required for a managed compute environment. For more information, see Compute Environments in the Batch User Guide.

updateComputeEnvironment_serviceRole :: Lens' UpdateComputeEnvironment (Maybe Text)

The full Amazon Resource Name (ARN) of the IAM role that allows Batch to make calls to other Amazon Web Services services on your behalf. For more information, see Batch service IAM role in the Batch User Guide.

If the compute environment has a service-linked role, it can't be changed to use a regular IAM role. Likewise, if the compute environment has a regular IAM role, it can't be changed to use a service-linked role.

If your specified role has a path other than /, then you must either specify the full role ARN (this is recommended) or prefix the role name with the path.

Depending on how you created your Batch service role, its ARN might contain the service-role path prefix. When you only specify the name of the service role, Batch assumes that your ARN doesn't use the service-role path prefix. Because of this, we recommend that you specify the full ARN of your service role when you create compute environments.

updateComputeEnvironment_computeEnvironment :: Lens' UpdateComputeEnvironment Text

The name or full Amazon Resource Name (ARN) of the compute environment to update.

Destructuring the Response

data UpdateComputeEnvironmentResponse

See: newUpdateComputeEnvironmentResponse smart constructor.

Constructors

UpdateComputeEnvironmentResponse'

Fields

computeEnvironmentName :: Maybe Text

The name of the compute environment. Up to 128 letters (uppercase and lowercase), numbers, hyphens, and underscores are allowed.

computeEnvironmentArn :: Maybe Text

The Amazon Resource Name (ARN) of the compute environment.

httpStatus :: Int

The response's http status code.

newUpdateComputeEnvironmentResponse

Arguments

:: Int
$sel:httpStatus:UpdateComputeEnvironmentResponse'

-> UpdateComputeEnvironmentResponse

Create a value of UpdateComputeEnvironmentResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:computeEnvironmentName:UpdateComputeEnvironmentResponse', updateComputeEnvironmentResponse_computeEnvironmentName - The name of the compute environment. Up to 128 letters (uppercase and lowercase), numbers, hyphens, and underscores are allowed.

$sel:computeEnvironmentArn:UpdateComputeEnvironmentResponse', updateComputeEnvironmentResponse_computeEnvironmentArn - The Amazon Resource Name (ARN) of the compute environment.

$sel:httpStatus:UpdateComputeEnvironmentResponse', updateComputeEnvironmentResponse_httpStatus - The response's http status code.

Response Lenses

updateComputeEnvironmentResponse_computeEnvironmentName :: Lens' UpdateComputeEnvironmentResponse (Maybe Text)

The name of the compute environment. Up to 128 letters (uppercase and lowercase), numbers, hyphens, and underscores are allowed.

updateComputeEnvironmentResponse_computeEnvironmentArn :: Lens' UpdateComputeEnvironmentResponse (Maybe Text)

The Amazon Resource Name (ARN) of the compute environment.

updateComputeEnvironmentResponse_httpStatus :: Lens' UpdateComputeEnvironmentResponse Int

The response's http status code.