Amazonka.Braket

Description

Derived from API version 2019-09-01 of the AWS service descriptions, licensed under Apache 2.0.

The Amazon Braket API Reference provides information about the operations and structures supported in Amazon Braket.

Service Configuration

defaultService :: Service

API version 2019-09-01 of the Amazon Braket SDK configuration.

Errors

Error matchers are designed for use with the functions provided by Control.Exception.Lens. This allows catching (and rethrowing) service specific errors returned by Braket.

ValidationException

_ValidationException :: AsError a => Getting (First ServiceError) a ServiceError

The input fails to satisfy the constraints specified by an AWS service.

AccessDeniedException

_AccessDeniedException :: AsError a => Getting (First ServiceError) a ServiceError

You do not have sufficient access to perform this action.

DeviceOfflineException

_DeviceOfflineException :: AsError a => Getting (First ServiceError) a ServiceError

The specified device is currently offline.

DeviceRetiredException

_DeviceRetiredException :: AsError a => Getting (First ServiceError) a ServiceError

The specified device has been retired.

ConflictException

_ConflictException :: AsError a => Getting (First ServiceError) a ServiceError

An error occurred due to a conflict.

ServiceQuotaExceededException

_ServiceQuotaExceededException :: AsError a => Getting (First ServiceError) a ServiceError

The request failed because a service quota is exceeded.

ThrottlingException

_ThrottlingException :: AsError a => Getting (First ServiceError) a ServiceError

The throttling rate limit is met.

InternalServiceException

_InternalServiceException :: AsError a => Getting (First ServiceError) a ServiceError

The request processing has failed because of an unknown error, exception, or failure.

ResourceNotFoundException

_ResourceNotFoundException :: AsError a => Getting (First ServiceError) a ServiceError

The specified resource was not found.

Waiters

Waiters poll by repeatedly sending a request until some remote success condition configured by the Wait specification is fulfilled. The Wait specification determines how many attempts should be made, in addition to delay and retry strategies.

Operations

Some AWS operations return results that are incomplete and require subsequent requests in order to obtain the entire result set. The process of sending subsequent requests to continue where a previous request left off is called pagination. For example, the ListObjects operation of Amazon S3 returns up to 1000 objects at a time, and you must send subsequent requests with the appropriate Marker in order to retrieve the next page of results.

Operations that have an AWSPager instance can transparently perform subsequent requests, correctly setting Markers and other request facets to iterate through the entire result set of a truncated API operation. Operations which support this have an additional note in the documentation.

Many operations have the ability to filter results on the server side. See the individual operation parameters for details.

GetQuantumTask

data GetQuantumTask

See: newGetQuantumTask smart constructor.

Constructors

GetQuantumTask' Text

newGetQuantumTask

Arguments

:: Text
$sel:quantumTaskArn:GetQuantumTask'

-> GetQuantumTask

Create a value of GetQuantumTask with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:quantumTaskArn:GetQuantumTask', getQuantumTask_quantumTaskArn - the ARN of the task to retrieve.

data GetQuantumTaskResponse

See: newGetQuantumTaskResponse smart constructor.

Constructors

GetQuantumTaskResponse' (Maybe Text) (Maybe POSIX) (Maybe (HashMap Text Text)) Int POSIX Text Text Text Text Text Integer QuantumTaskStatus

newGetQuantumTaskResponse

Arguments

:: Int
$sel:httpStatus:GetQuantumTaskResponse'

-> UTCTime
$sel:createdAt:GetQuantumTaskResponse'

-> Text
$sel:deviceArn:GetQuantumTaskResponse'

-> Text
$sel:deviceParameters:GetQuantumTaskResponse'

-> Text
$sel:outputS3Bucket:GetQuantumTaskResponse'

-> Text
$sel:outputS3Directory:GetQuantumTaskResponse'

-> Text
$sel:quantumTaskArn:GetQuantumTask'

-> Integer
$sel:shots:GetQuantumTaskResponse'

-> QuantumTaskStatus
$sel:status:GetQuantumTaskResponse'

-> GetQuantumTaskResponse

Create a value of GetQuantumTaskResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:failureReason:GetQuantumTaskResponse', getQuantumTaskResponse_failureReason - The reason that a task failed.

$sel:endedAt:GetQuantumTaskResponse', getQuantumTaskResponse_endedAt - The time at which the task ended.

$sel:tags:GetQuantumTaskResponse', getQuantumTaskResponse_tags - The tags that belong to this task.

$sel:httpStatus:GetQuantumTaskResponse', getQuantumTaskResponse_httpStatus - The response's http status code.

$sel:createdAt:GetQuantumTaskResponse', getQuantumTaskResponse_createdAt - The time at which the task was created.

$sel:deviceArn:GetQuantumTaskResponse', getQuantumTaskResponse_deviceArn - The ARN of the device the task was run on.

$sel:deviceParameters:GetQuantumTaskResponse', getQuantumTaskResponse_deviceParameters - The parameters for the device on which the task ran.

$sel:outputS3Bucket:GetQuantumTaskResponse', getQuantumTaskResponse_outputS3Bucket - The S3 bucket where task results are stored.

$sel:outputS3Directory:GetQuantumTaskResponse', getQuantumTaskResponse_outputS3Directory - The folder in the S3 bucket where task results are stored.

$sel:quantumTaskArn:GetQuantumTask', getQuantumTaskResponse_quantumTaskArn - The ARN of the task.

$sel:shots:GetQuantumTaskResponse', getQuantumTaskResponse_shots - The number of shots used in the task.

$sel:status:GetQuantumTaskResponse', getQuantumTaskResponse_status - The status of the task.

SearchQuantumTasks (Paginated)

data SearchQuantumTasks

See: newSearchQuantumTasks smart constructor.

Constructors

SearchQuantumTasks' (Maybe Text) (Maybe Natural) [SearchQuantumTasksFilter]

newSearchQuantumTasks :: SearchQuantumTasks

Create a value of SearchQuantumTasks with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:nextToken:SearchQuantumTasks', searchQuantumTasks_nextToken - A token used for pagination of results returned in the response. Use the token returned from the previous request continue results where the previous request ended.

$sel:maxResults:SearchQuantumTasks', searchQuantumTasks_maxResults - Maximum number of results to return in the response.

$sel:filters:SearchQuantumTasks', searchQuantumTasks_filters - Array of SearchQuantumTasksFilter objects.

data SearchQuantumTasksResponse

See: newSearchQuantumTasksResponse smart constructor.

Constructors

SearchQuantumTasksResponse' (Maybe Text) Int [QuantumTaskSummary]

newSearchQuantumTasksResponse

Arguments

:: Int
$sel:httpStatus:SearchQuantumTasksResponse'

-> SearchQuantumTasksResponse

Create a value of SearchQuantumTasksResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:nextToken:SearchQuantumTasks', searchQuantumTasksResponse_nextToken - A token used for pagination of results, or null if there are no additional results. Use the token value in a subsequent request to continue results where the previous request ended.

$sel:httpStatus:SearchQuantumTasksResponse', searchQuantumTasksResponse_httpStatus - The response's http status code.

$sel:quantumTasks:SearchQuantumTasksResponse', searchQuantumTasksResponse_quantumTasks - An array of QuantumTaskSummary objects for tasks that match the specified filters.

ListTagsForResource

data ListTagsForResource

See: newListTagsForResource smart constructor.

Constructors

ListTagsForResource' Text

newListTagsForResource

Arguments

:: Text
$sel:resourceArn:ListTagsForResource'

-> ListTagsForResource

Create a value of ListTagsForResource with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:resourceArn:ListTagsForResource', listTagsForResource_resourceArn - Specify the resourceArn for the resource whose tags to display.

data ListTagsForResourceResponse

See: newListTagsForResourceResponse smart constructor.

Constructors

ListTagsForResourceResponse' (Maybe (HashMap Text Text)) Int

newListTagsForResourceResponse

Arguments

:: Int
$sel:httpStatus:ListTagsForResourceResponse'

-> ListTagsForResourceResponse

Create a value of ListTagsForResourceResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:tags:ListTagsForResourceResponse', listTagsForResourceResponse_tags - Displays the key, value pairs of tags associated with this resource.

$sel:httpStatus:ListTagsForResourceResponse', listTagsForResourceResponse_httpStatus - The response's http status code.

GetDevice

data GetDevice

See: newGetDevice smart constructor.

Constructors

GetDevice' Text

newGetDevice

Arguments

:: Text
$sel:deviceArn:GetDevice'

-> GetDevice

Create a value of GetDevice with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:deviceArn:GetDevice', getDevice_deviceArn - The ARN of the device to retrieve.

data GetDeviceResponse

See: newGetDeviceResponse smart constructor.

Constructors

GetDeviceResponse' Int Text Text Text DeviceStatus DeviceType Text

newGetDeviceResponse

Arguments

:: Int
$sel:httpStatus:GetDeviceResponse'

-> Text
$sel:deviceArn:GetDevice'

-> Text
$sel:deviceCapabilities:GetDeviceResponse'

-> Text
$sel:deviceName:GetDeviceResponse'

-> DeviceStatus
$sel:deviceStatus:GetDeviceResponse'

-> DeviceType
$sel:deviceType:GetDeviceResponse'

-> Text
$sel:providerName:GetDeviceResponse'

-> GetDeviceResponse

Create a value of GetDeviceResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:httpStatus:GetDeviceResponse', getDeviceResponse_httpStatus - The response's http status code.

$sel:deviceArn:GetDevice', getDeviceResponse_deviceArn - The ARN of the device.

$sel:deviceCapabilities:GetDeviceResponse', getDeviceResponse_deviceCapabilities - Details about the capabilities of the device.

$sel:deviceName:GetDeviceResponse', getDeviceResponse_deviceName - The name of the device.

$sel:deviceStatus:GetDeviceResponse', getDeviceResponse_deviceStatus - The status of the device.

$sel:deviceType:GetDeviceResponse', getDeviceResponse_deviceType - The type of the device.

$sel:providerName:GetDeviceResponse', getDeviceResponse_providerName - The name of the partner company for the device.

CancelQuantumTask

data CancelQuantumTask

See: newCancelQuantumTask smart constructor.

Constructors

CancelQuantumTask' Text Text

newCancelQuantumTask

Arguments

:: Text
$sel:clientToken:CancelQuantumTask'

-> Text
$sel:quantumTaskArn:CancelQuantumTask'

-> CancelQuantumTask

Create a value of CancelQuantumTask with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:clientToken:CancelQuantumTask', cancelQuantumTask_clientToken - The client token associated with the request.

$sel:quantumTaskArn:CancelQuantumTask', cancelQuantumTask_quantumTaskArn - The ARN of the task to cancel.

data CancelQuantumTaskResponse

See: newCancelQuantumTaskResponse smart constructor.

Constructors

CancelQuantumTaskResponse' Int CancellationStatus Text

newCancelQuantumTaskResponse

Arguments

:: Int
$sel:httpStatus:CancelQuantumTaskResponse'

-> CancellationStatus
$sel:cancellationStatus:CancelQuantumTaskResponse'

-> Text
$sel:quantumTaskArn:CancelQuantumTask'

-> CancelQuantumTaskResponse

Create a value of CancelQuantumTaskResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:httpStatus:CancelQuantumTaskResponse', cancelQuantumTaskResponse_httpStatus - The response's http status code.

$sel:cancellationStatus:CancelQuantumTaskResponse', cancelQuantumTaskResponse_cancellationStatus - The status of the cancellation request.

$sel:quantumTaskArn:CancelQuantumTask', cancelQuantumTaskResponse_quantumTaskArn - The ARN of the task.

SearchDevices (Paginated)

data SearchDevices

See: newSearchDevices smart constructor.

Constructors

SearchDevices' (Maybe Text) (Maybe Natural) [SearchDevicesFilter]

newSearchDevices :: SearchDevices

Create a value of SearchDevices with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:nextToken:SearchDevices', searchDevices_nextToken - A token used for pagination of results returned in the response. Use the token returned from the previous request continue results where the previous request ended.

$sel:maxResults:SearchDevices', searchDevices_maxResults - The maximum number of results to return in the response.

$sel:filters:SearchDevices', searchDevices_filters - The filter values to use to search for a device.

data SearchDevicesResponse

See: newSearchDevicesResponse smart constructor.

Constructors

SearchDevicesResponse' (Maybe Text) Int [DeviceSummary]

newSearchDevicesResponse

Arguments

:: Int
$sel:httpStatus:SearchDevicesResponse'

-> SearchDevicesResponse

Create a value of SearchDevicesResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:nextToken:SearchDevices', searchDevicesResponse_nextToken - A token used for pagination of results, or null if there are no additional results. Use the token value in a subsequent request to continue results where the previous request ended.

$sel:httpStatus:SearchDevicesResponse', searchDevicesResponse_httpStatus - The response's http status code.

$sel:devices:SearchDevicesResponse', searchDevicesResponse_devices - An array of DeviceSummary objects for devices that match the specified filter values.

TagResource

data TagResource

See: newTagResource smart constructor.

Constructors

TagResource' Text (HashMap Text Text)

newTagResource

Arguments

:: Text
$sel:resourceArn:TagResource'

-> TagResource

Create a value of TagResource with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:resourceArn:TagResource', tagResource_resourceArn - Specify the resourceArn of the resource to which a tag will be added.

$sel:tags:TagResource', tagResource_tags - Specify the tags to add to the resource.

data TagResourceResponse

See: newTagResourceResponse smart constructor.

Constructors

TagResourceResponse' Int

newTagResourceResponse

Arguments

:: Int
$sel:httpStatus:TagResourceResponse'

-> TagResourceResponse

Create a value of TagResourceResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:httpStatus:TagResourceResponse', tagResourceResponse_httpStatus - The response's http status code.

UntagResource

data UntagResource

See: newUntagResource smart constructor.

Constructors

UntagResource' Text [Text]

newUntagResource

Arguments

:: Text
$sel:resourceArn:UntagResource'

-> UntagResource

Create a value of UntagResource with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:resourceArn:UntagResource', untagResource_resourceArn - Specify the resourceArn for the resource from which to remove the tags.

$sel:tagKeys:UntagResource', untagResource_tagKeys - Specify the keys for the tags to remove from the resource.

data UntagResourceResponse

See: newUntagResourceResponse smart constructor.

Constructors

UntagResourceResponse' Int

newUntagResourceResponse

Arguments

:: Int
$sel:httpStatus:UntagResourceResponse'

-> UntagResourceResponse

Create a value of UntagResourceResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:httpStatus:UntagResourceResponse', untagResourceResponse_httpStatus - The response's http status code.

CreateQuantumTask

data CreateQuantumTask

See: newCreateQuantumTask smart constructor.

Constructors

CreateQuantumTask' (Maybe Text) (Maybe (HashMap Text Text)) Text Text Text Text Text Natural

newCreateQuantumTask

Arguments

:: Text
$sel:action:CreateQuantumTask'

-> Text
$sel:clientToken:CreateQuantumTask'

-> Text
$sel:deviceArn:CreateQuantumTask'

-> Text
$sel:outputS3Bucket:CreateQuantumTask'

-> Text
$sel:outputS3KeyPrefix:CreateQuantumTask'

-> Natural
$sel:shots:CreateQuantumTask'

-> CreateQuantumTask

Create a value of CreateQuantumTask with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:deviceParameters:CreateQuantumTask', createQuantumTask_deviceParameters - The parameters for the device to run the task on.

$sel:tags:CreateQuantumTask', createQuantumTask_tags - Tags to be added to the quantum task you're creating.

$sel:action:CreateQuantumTask', createQuantumTask_action - The action associated with the task.

$sel:clientToken:CreateQuantumTask', createQuantumTask_clientToken - The client token associated with the request.

$sel:deviceArn:CreateQuantumTask', createQuantumTask_deviceArn - The ARN of the device to run the task on.

$sel:outputS3Bucket:CreateQuantumTask', createQuantumTask_outputS3Bucket - The S3 bucket to store task result files in.

$sel:outputS3KeyPrefix:CreateQuantumTask', createQuantumTask_outputS3KeyPrefix - The key prefix for the location in the S3 bucket to store task results in.

$sel:shots:CreateQuantumTask', createQuantumTask_shots - The number of shots to use for the task.

data CreateQuantumTaskResponse

See: newCreateQuantumTaskResponse smart constructor.

Constructors

CreateQuantumTaskResponse' Int Text

newCreateQuantumTaskResponse

Arguments

:: Int
$sel:httpStatus:CreateQuantumTaskResponse'

-> Text
$sel:quantumTaskArn:CreateQuantumTaskResponse'

-> CreateQuantumTaskResponse

Create a value of CreateQuantumTaskResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:httpStatus:CreateQuantumTaskResponse', createQuantumTaskResponse_httpStatus - The response's http status code.

$sel:quantumTaskArn:CreateQuantumTaskResponse', createQuantumTaskResponse_quantumTaskArn - The ARN of the task created by the request.

Types

CancellationStatus

newtype CancellationStatus

Constructors

CancellationStatus'

Fields

fromCancellationStatus :: Text

Bundled Patterns

pattern CancellationStatus_CANCELLED :: CancellationStatus

pattern CancellationStatus_CANCELLING :: CancellationStatus

DeviceStatus

newtype DeviceStatus

Constructors

DeviceStatus'

Fields

fromDeviceStatus :: Text

Bundled Patterns

pattern DeviceStatus_OFFLINE :: DeviceStatus

pattern DeviceStatus_ONLINE :: DeviceStatus

pattern DeviceStatus_RETIRED :: DeviceStatus

DeviceType

newtype DeviceType

Constructors

DeviceType'

Fields

fromDeviceType :: Text

Bundled Patterns

pattern DeviceType_QPU :: DeviceType

pattern DeviceType_SIMULATOR :: DeviceType

QuantumTaskStatus

newtype QuantumTaskStatus

Constructors

QuantumTaskStatus'

Fields

fromQuantumTaskStatus :: Text

Bundled Patterns

pattern QuantumTaskStatus_CANCELLED :: QuantumTaskStatus

pattern QuantumTaskStatus_CANCELLING :: QuantumTaskStatus

pattern QuantumTaskStatus_COMPLETED :: QuantumTaskStatus

pattern QuantumTaskStatus_CREATED :: QuantumTaskStatus

pattern QuantumTaskStatus_FAILED :: QuantumTaskStatus

pattern QuantumTaskStatus_QUEUED :: QuantumTaskStatus

pattern QuantumTaskStatus_RUNNING :: QuantumTaskStatus

SearchQuantumTasksFilterOperator

newtype SearchQuantumTasksFilterOperator

Constructors

SearchQuantumTasksFilterOperator'

Fields

fromSearchQuantumTasksFilterOperator :: Text

Bundled Patterns

pattern SearchQuantumTasksFilterOperator_BETWEEN :: SearchQuantumTasksFilterOperator

pattern SearchQuantumTasksFilterOperator_EQUAL :: SearchQuantumTasksFilterOperator

pattern SearchQuantumTasksFilterOperator_GT :: SearchQuantumTasksFilterOperator

pattern SearchQuantumTasksFilterOperator_GTE :: SearchQuantumTasksFilterOperator

pattern SearchQuantumTasksFilterOperator_LT :: SearchQuantumTasksFilterOperator

pattern SearchQuantumTasksFilterOperator_LTE :: SearchQuantumTasksFilterOperator

DeviceSummary

data DeviceSummary

Includes information about the device.

See: newDeviceSummary smart constructor.

Constructors

DeviceSummary' Text Text DeviceStatus DeviceType Text

newDeviceSummary

Arguments

:: Text
$sel:deviceArn:DeviceSummary'

-> Text
$sel:deviceName:DeviceSummary'

-> DeviceStatus
$sel:deviceStatus:DeviceSummary'

-> DeviceType
$sel:deviceType:DeviceSummary'

-> Text
$sel:providerName:DeviceSummary'

-> DeviceSummary

Create a value of DeviceSummary with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:deviceArn:DeviceSummary', deviceSummary_deviceArn - The ARN of the device.

$sel:deviceName:DeviceSummary', deviceSummary_deviceName - The name of the device.

$sel:deviceStatus:DeviceSummary', deviceSummary_deviceStatus - The status of the device.

$sel:deviceType:DeviceSummary', deviceSummary_deviceType - The type of the device.

$sel:providerName:DeviceSummary', deviceSummary_providerName - The provider of the device.

QuantumTaskSummary

data QuantumTaskSummary

Includes information about a quantum task.

See: newQuantumTaskSummary smart constructor.

Constructors

QuantumTaskSummary' (Maybe POSIX) (Maybe (HashMap Text Text)) POSIX Text Text Text Text Integer QuantumTaskStatus

newQuantumTaskSummary

Arguments

:: UTCTime
$sel:createdAt:QuantumTaskSummary'

-> Text
$sel:deviceArn:QuantumTaskSummary'

-> Text
$sel:outputS3Bucket:QuantumTaskSummary'

-> Text
$sel:outputS3Directory:QuantumTaskSummary'

-> Text
$sel:quantumTaskArn:QuantumTaskSummary'

-> Integer
$sel:shots:QuantumTaskSummary'

-> QuantumTaskStatus
$sel:status:QuantumTaskSummary'

-> QuantumTaskSummary

Create a value of QuantumTaskSummary with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:endedAt:QuantumTaskSummary', quantumTaskSummary_endedAt - The time at which the task finished.

$sel:tags:QuantumTaskSummary', quantumTaskSummary_tags - Displays the key, value pairs of tags associated with this quantum task.

$sel:createdAt:QuantumTaskSummary', quantumTaskSummary_createdAt - The time at which the task was created.

$sel:deviceArn:QuantumTaskSummary', quantumTaskSummary_deviceArn - The ARN of the device the task ran on.

$sel:outputS3Bucket:QuantumTaskSummary', quantumTaskSummary_outputS3Bucket - The S3 bucket where the task result file is stored..

$sel:outputS3Directory:QuantumTaskSummary', quantumTaskSummary_outputS3Directory - The folder in the S3 bucket where the task result file is stored.

$sel:quantumTaskArn:QuantumTaskSummary', quantumTaskSummary_quantumTaskArn - The ARN of the task.

$sel:shots:QuantumTaskSummary', quantumTaskSummary_shots - The shots used for the task.

$sel:status:QuantumTaskSummary', quantumTaskSummary_status - The status of the task.

SearchDevicesFilter

data SearchDevicesFilter

The filter to use for searching devices.

See: newSearchDevicesFilter smart constructor.

Constructors

SearchDevicesFilter' Text (NonEmpty Text)

newSearchDevicesFilter

Arguments

:: Text
$sel:name:SearchDevicesFilter'

-> NonEmpty Text
$sel:values:SearchDevicesFilter'

-> SearchDevicesFilter

Create a value of SearchDevicesFilter with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:name:SearchDevicesFilter', searchDevicesFilter_name - The name to use to filter results.

$sel:values:SearchDevicesFilter', searchDevicesFilter_values - The values to use to filter results.

SearchQuantumTasksFilter

data SearchQuantumTasksFilter

A filter to use to search for tasks.

See: newSearchQuantumTasksFilter smart constructor.

Constructors

SearchQuantumTasksFilter' Text SearchQuantumTasksFilterOperator (NonEmpty Text)

newSearchQuantumTasksFilter

Arguments

:: Text
$sel:name:SearchQuantumTasksFilter'

-> SearchQuantumTasksFilterOperator
$sel:operator:SearchQuantumTasksFilter'

-> NonEmpty Text
$sel:values:SearchQuantumTasksFilter'

-> SearchQuantumTasksFilter

Create a value of SearchQuantumTasksFilter with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:name:SearchQuantumTasksFilter', searchQuantumTasksFilter_name - The name of the device used for the task.

$sel:operator:SearchQuantumTasksFilter', searchQuantumTasksFilter_operator - An operator to use in the filter.

$sel:values:SearchQuantumTasksFilter', searchQuantumTasksFilter_values - The values to use for the filter.