Amazonka.Braket.GetQuantumTask

Description

Retrieves the specified quantum task.

Creating a Request

data GetQuantumTask

See: newGetQuantumTask smart constructor.

Constructors

GetQuantumTask'

Fields

quantumTaskArn :: Text

the ARN of the task to retrieve.

newGetQuantumTask

Arguments

:: Text
$sel:quantumTaskArn:GetQuantumTask'

-> GetQuantumTask

Create a value of GetQuantumTask with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:quantumTaskArn:GetQuantumTask', getQuantumTask_quantumTaskArn - the ARN of the task to retrieve.

Request Lenses

getQuantumTask_quantumTaskArn :: Lens' GetQuantumTask Text

the ARN of the task to retrieve.

Destructuring the Response

data GetQuantumTaskResponse

See: newGetQuantumTaskResponse smart constructor.

Constructors

GetQuantumTaskResponse'

Fields

failureReason :: Maybe Text

The reason that a task failed.

endedAt :: Maybe POSIX

The time at which the task ended.

tags :: Maybe (HashMap Text Text)

The tags that belong to this task.

httpStatus :: Int

The response's http status code.

createdAt :: POSIX

The time at which the task was created.

deviceArn :: Text

The ARN of the device the task was run on.

deviceParameters :: Text

The parameters for the device on which the task ran.

outputS3Bucket :: Text

The S3 bucket where task results are stored.

outputS3Directory :: Text

The folder in the S3 bucket where task results are stored.

quantumTaskArn :: Text

The ARN of the task.

shots :: Integer

The number of shots used in the task.

status :: QuantumTaskStatus

The status of the task.

newGetQuantumTaskResponse

Arguments

:: Int
$sel:httpStatus:GetQuantumTaskResponse'

-> UTCTime
$sel:createdAt:GetQuantumTaskResponse'

-> Text
$sel:deviceArn:GetQuantumTaskResponse'

-> Text
$sel:deviceParameters:GetQuantumTaskResponse'

-> Text
$sel:outputS3Bucket:GetQuantumTaskResponse'

-> Text
$sel:outputS3Directory:GetQuantumTaskResponse'

-> Text
$sel:quantumTaskArn:GetQuantumTask'

-> Integer
$sel:shots:GetQuantumTaskResponse'

-> QuantumTaskStatus
$sel:status:GetQuantumTaskResponse'

-> GetQuantumTaskResponse

Create a value of GetQuantumTaskResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:failureReason:GetQuantumTaskResponse', getQuantumTaskResponse_failureReason - The reason that a task failed.

$sel:endedAt:GetQuantumTaskResponse', getQuantumTaskResponse_endedAt - The time at which the task ended.

$sel:tags:GetQuantumTaskResponse', getQuantumTaskResponse_tags - The tags that belong to this task.

$sel:httpStatus:GetQuantumTaskResponse', getQuantumTaskResponse_httpStatus - The response's http status code.

$sel:createdAt:GetQuantumTaskResponse', getQuantumTaskResponse_createdAt - The time at which the task was created.

$sel:deviceArn:GetQuantumTaskResponse', getQuantumTaskResponse_deviceArn - The ARN of the device the task was run on.

$sel:deviceParameters:GetQuantumTaskResponse', getQuantumTaskResponse_deviceParameters - The parameters for the device on which the task ran.

$sel:outputS3Bucket:GetQuantumTaskResponse', getQuantumTaskResponse_outputS3Bucket - The S3 bucket where task results are stored.

$sel:outputS3Directory:GetQuantumTaskResponse', getQuantumTaskResponse_outputS3Directory - The folder in the S3 bucket where task results are stored.

$sel:quantumTaskArn:GetQuantumTask', getQuantumTaskResponse_quantumTaskArn - The ARN of the task.

$sel:shots:GetQuantumTaskResponse', getQuantumTaskResponse_shots - The number of shots used in the task.

$sel:status:GetQuantumTaskResponse', getQuantumTaskResponse_status - The status of the task.

Response Lenses

getQuantumTaskResponse_failureReason :: Lens' GetQuantumTaskResponse (Maybe Text)

The reason that a task failed.

getQuantumTaskResponse_endedAt :: Lens' GetQuantumTaskResponse (Maybe UTCTime)

The time at which the task ended.

getQuantumTaskResponse_tags :: Lens' GetQuantumTaskResponse (Maybe (HashMap Text Text))

The tags that belong to this task.

getQuantumTaskResponse_httpStatus :: Lens' GetQuantumTaskResponse Int

The response's http status code.

getQuantumTaskResponse_createdAt :: Lens' GetQuantumTaskResponse UTCTime

The time at which the task was created.

getQuantumTaskResponse_deviceArn :: Lens' GetQuantumTaskResponse Text

The ARN of the device the task was run on.

getQuantumTaskResponse_deviceParameters :: Lens' GetQuantumTaskResponse Text

The parameters for the device on which the task ran.

getQuantumTaskResponse_outputS3Bucket :: Lens' GetQuantumTaskResponse Text

The S3 bucket where task results are stored.

getQuantumTaskResponse_outputS3Directory :: Lens' GetQuantumTaskResponse Text

The folder in the S3 bucket where task results are stored.

getQuantumTaskResponse_quantumTaskArn :: Lens' GetQuantumTaskResponse Text

The ARN of the task.

getQuantumTaskResponse_shots :: Lens' GetQuantumTaskResponse Integer

The number of shots used in the task.

getQuantumTaskResponse_status :: Lens' GetQuantumTaskResponse QuantumTaskStatus

The status of the task.