Amazonka.Budgets.CreateNotification

Description

Creates a notification. You must create the budget before you create the associated notification.

Creating a Request

data CreateNotification

Request of CreateNotification

See: newCreateNotification smart constructor.

Constructors

CreateNotification'

Fields

accountId :: Text

The accountId that is associated with the budget that you want to create a notification for.

budgetName :: Text

The name of the budget that you want AWS to notify you about. Budget names must be unique within an account.

notification :: Notification

The notification that you want to create.

subscribers :: NonEmpty Subscriber

A list of subscribers that you want to associate with the notification. Each notification can have one SNS subscriber and up to 10 email subscribers.

newCreateNotification

Arguments

:: Text
$sel:accountId:CreateNotification'

-> Text
$sel:budgetName:CreateNotification'

-> Notification
$sel:notification:CreateNotification'

-> NonEmpty Subscriber
$sel:subscribers:CreateNotification'

-> CreateNotification

Create a value of CreateNotification with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:accountId:CreateNotification', createNotification_accountId - The accountId that is associated with the budget that you want to create a notification for.

$sel:budgetName:CreateNotification', createNotification_budgetName - The name of the budget that you want AWS to notify you about. Budget names must be unique within an account.

$sel:notification:CreateNotification', createNotification_notification - The notification that you want to create.

$sel:subscribers:CreateNotification', createNotification_subscribers - A list of subscribers that you want to associate with the notification. Each notification can have one SNS subscriber and up to 10 email subscribers.

Request Lenses

createNotification_accountId :: Lens' CreateNotification Text

The accountId that is associated with the budget that you want to create a notification for.

createNotification_budgetName :: Lens' CreateNotification Text

The name of the budget that you want AWS to notify you about. Budget names must be unique within an account.

createNotification_notification :: Lens' CreateNotification Notification

The notification that you want to create.

createNotification_subscribers :: Lens' CreateNotification (NonEmpty Subscriber)

A list of subscribers that you want to associate with the notification. Each notification can have one SNS subscriber and up to 10 email subscribers.

Destructuring the Response

data CreateNotificationResponse

Response of CreateNotification

See: newCreateNotificationResponse smart constructor.

Constructors

CreateNotificationResponse'

Fields

httpStatus :: Int

The response's http status code.

newCreateNotificationResponse

Arguments

:: Int
$sel:httpStatus:CreateNotificationResponse'

-> CreateNotificationResponse

Create a value of CreateNotificationResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:httpStatus:CreateNotificationResponse', createNotificationResponse_httpStatus - The response's http status code.

Response Lenses

createNotificationResponse_httpStatus :: Lens' CreateNotificationResponse Int

The response's http status code.