Amazonka.Budgets.Types.Notification

Description

Documentation

data Notification

A notification that is associated with a budget. A budget can have up to ten notifications.

Each notification must have at least one subscriber. A notification can have one SNS subscriber and up to 10 email subscribers, for a total of 11 subscribers.

For example, if you have a budget for 200 dollars and you want to be notified when you go over 160 dollars, create a notification with the following parameters:

A notificationType of ACTUAL

A thresholdType of PERCENTAGE

A comparisonOperator of GREATER_THAN

A notification threshold of 80

See: newNotification smart constructor.

Constructors

Notification'

Fields

thresholdType :: Maybe ThresholdType

The type of threshold for a notification. For ABSOLUTE_VALUE thresholds, AWS notifies you when you go over or are forecasted to go over your total cost threshold. For PERCENTAGE thresholds, AWS notifies you when you go over or are forecasted to go over a certain percentage of your forecasted spend. For example, if you have a budget for 200 dollars and you have a PERCENTAGE threshold of 80%, AWS notifies you when you go over 160 dollars.

notificationState :: Maybe NotificationState

Whether this notification is in alarm. If a budget notification is in the ALARM state, you have passed the set threshold for the budget.

notificationType :: NotificationType

Whether the notification is for how much you have spent (ACTUAL) or for how much you're forecasted to spend (FORECASTED).

comparisonOperator :: ComparisonOperator

The comparison that is used for this notification.

threshold :: Double

The threshold that is associated with a notification. Thresholds are always a percentage, and many customers find value being alerted between 50% - 200% of the budgeted amount. The maximum limit for your threshold is 1,000,000% above the budgeted amount.

newNotification

Arguments

:: NotificationType
$sel:notificationType:Notification'

-> ComparisonOperator
$sel:comparisonOperator:Notification'

-> Double
$sel:threshold:Notification'

-> Notification

Create a value of Notification with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:thresholdType:Notification', notification_thresholdType - The type of threshold for a notification. For ABSOLUTE_VALUE thresholds, AWS notifies you when you go over or are forecasted to go over your total cost threshold. For PERCENTAGE thresholds, AWS notifies you when you go over or are forecasted to go over a certain percentage of your forecasted spend. For example, if you have a budget for 200 dollars and you have a PERCENTAGE threshold of 80%, AWS notifies you when you go over 160 dollars.

$sel:notificationState:Notification', notification_notificationState - Whether this notification is in alarm. If a budget notification is in the ALARM state, you have passed the set threshold for the budget.

$sel:notificationType:Notification', notification_notificationType - Whether the notification is for how much you have spent (ACTUAL) or for how much you're forecasted to spend (FORECASTED).

$sel:comparisonOperator:Notification', notification_comparisonOperator - The comparison that is used for this notification.

$sel:threshold:Notification', notification_threshold - The threshold that is associated with a notification. Thresholds are always a percentage, and many customers find value being alerted between 50% - 200% of the budgeted amount. The maximum limit for your threshold is 1,000,000% above the budgeted amount.

notification_thresholdType :: Lens' Notification (Maybe ThresholdType)

The type of threshold for a notification. For ABSOLUTE_VALUE thresholds, AWS notifies you when you go over or are forecasted to go over your total cost threshold. For PERCENTAGE thresholds, AWS notifies you when you go over or are forecasted to go over a certain percentage of your forecasted spend. For example, if you have a budget for 200 dollars and you have a PERCENTAGE threshold of 80%, AWS notifies you when you go over 160 dollars.

notification_notificationState :: Lens' Notification (Maybe NotificationState)

Whether this notification is in alarm. If a budget notification is in the ALARM state, you have passed the set threshold for the budget.

notification_notificationType :: Lens' Notification NotificationType

Whether the notification is for how much you have spent (ACTUAL) or for how much you're forecasted to spend (FORECASTED).

notification_comparisonOperator :: Lens' Notification ComparisonOperator

The comparison that is used for this notification.

notification_threshold :: Lens' Notification Double

The threshold that is associated with a notification. Thresholds are always a percentage, and many customers find value being alerted between 50% - 200% of the budgeted amount. The maximum limit for your threshold is 1,000,000% above the budgeted amount.