Amazonka.CertificateManagerPCA.CreateCertificateAuthorityAuditReport

Description

Creates an audit report that lists every time that your CA private key is used. The report is saved in the Amazon S3 bucket that you specify on input. The IssueCertificate and RevokeCertificate actions use the private key.

Both PCA and the IAM principal must have permission to write to the S3 bucket that you specify. If the IAM principal making the call does not have permission to write to the bucket, then an exception is thrown. For more information, see Configure Access to ACM Private CA.

ACM Private CA assets that are stored in Amazon S3 can be protected with encryption. For more information, see Encrypting Your Audit Reports.

Creating a Request

data CreateCertificateAuthorityAuditReport

See: newCreateCertificateAuthorityAuditReport smart constructor.

Constructors

CreateCertificateAuthorityAuditReport'

Fields

certificateAuthorityArn :: Text

The Amazon Resource Name (ARN) of the CA to be audited. This is of the form:

arn:aws:acm-pca:region:account:certificate-authority/12345678-1234-1234-1234-123456789012 .

s3BucketName :: Text

The name of the S3 bucket that will contain the audit report.

auditReportResponseFormat :: AuditReportResponseFormat

The format in which to create the report. This can be either JSON or CSV.

newCreateCertificateAuthorityAuditReport

Arguments

:: Text
$sel:certificateAuthorityArn:CreateCertificateAuthorityAuditReport'

-> Text
$sel:s3BucketName:CreateCertificateAuthorityAuditReport'

-> AuditReportResponseFormat
$sel:auditReportResponseFormat:CreateCertificateAuthorityAuditReport'

-> CreateCertificateAuthorityAuditReport

Create a value of CreateCertificateAuthorityAuditReport with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:certificateAuthorityArn:CreateCertificateAuthorityAuditReport', createCertificateAuthorityAuditReport_certificateAuthorityArn - The Amazon Resource Name (ARN) of the CA to be audited. This is of the form:

arn:aws:acm-pca:region:account:certificate-authority/12345678-1234-1234-1234-123456789012 .

$sel:s3BucketName:CreateCertificateAuthorityAuditReport', createCertificateAuthorityAuditReport_s3BucketName - The name of the S3 bucket that will contain the audit report.

$sel:auditReportResponseFormat:CreateCertificateAuthorityAuditReport', createCertificateAuthorityAuditReport_auditReportResponseFormat - The format in which to create the report. This can be either JSON or CSV.

Request Lenses

createCertificateAuthorityAuditReport_certificateAuthorityArn :: Lens' CreateCertificateAuthorityAuditReport Text

The Amazon Resource Name (ARN) of the CA to be audited. This is of the form:

arn:aws:acm-pca:region:account:certificate-authority/12345678-1234-1234-1234-123456789012 .

createCertificateAuthorityAuditReport_s3BucketName :: Lens' CreateCertificateAuthorityAuditReport Text

The name of the S3 bucket that will contain the audit report.

createCertificateAuthorityAuditReport_auditReportResponseFormat :: Lens' CreateCertificateAuthorityAuditReport AuditReportResponseFormat

The format in which to create the report. This can be either JSON or CSV.

Destructuring the Response

data CreateCertificateAuthorityAuditReportResponse

See: newCreateCertificateAuthorityAuditReportResponse smart constructor.

Constructors

CreateCertificateAuthorityAuditReportResponse'

Fields

s3Key :: Maybe Text

The key that uniquely identifies the report file in your S3 bucket.

auditReportId :: Maybe Text

An alphanumeric string that contains a report identifier.

httpStatus :: Int

The response's http status code.

newCreateCertificateAuthorityAuditReportResponse

Arguments

:: Int
$sel:httpStatus:CreateCertificateAuthorityAuditReportResponse'

-> CreateCertificateAuthorityAuditReportResponse

Create a value of CreateCertificateAuthorityAuditReportResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:s3Key:CreateCertificateAuthorityAuditReportResponse', createCertificateAuthorityAuditReportResponse_s3Key - The key that uniquely identifies the report file in your S3 bucket.

$sel:auditReportId:CreateCertificateAuthorityAuditReportResponse', createCertificateAuthorityAuditReportResponse_auditReportId - An alphanumeric string that contains a report identifier.

$sel:httpStatus:CreateCertificateAuthorityAuditReportResponse', createCertificateAuthorityAuditReportResponse_httpStatus - The response's http status code.

Response Lenses

createCertificateAuthorityAuditReportResponse_s3Key :: Lens' CreateCertificateAuthorityAuditReportResponse (Maybe Text)

The key that uniquely identifies the report file in your S3 bucket.

createCertificateAuthorityAuditReportResponse_auditReportId :: Lens' CreateCertificateAuthorityAuditReportResponse (Maybe Text)

An alphanumeric string that contains a report identifier.

createCertificateAuthorityAuditReportResponse_httpStatus :: Lens' CreateCertificateAuthorityAuditReportResponse Int

The response's http status code.