Amazonka.CertificateManagerPCA.CreatePermission

Description

Grants one or more permissions on a private CA to the AWS Certificate Manager (ACM) service principal (acm.amazonaws.com). These permissions allow ACM to issue and renew ACM certificates that reside in the same AWS account as the CA.

You can list current permissions with the ListPermissions action and revoke them with the DeletePermission action.

About Permissions

If the private CA and the certificates it issues reside in the same account, you can use CreatePermission to grant permissions for ACM to carry out automatic certificate renewals.

For automatic certificate renewal to succeed, the ACM service principal needs permissions to create, retrieve, and list certificates.

If the private CA and the ACM certificates reside in different accounts, then permissions cannot be used to enable automatic renewals. Instead, the ACM certificate owner must set up a resource-based policy to enable cross-account issuance and renewals. For more information, see Using a Resource Based Policy with ACM Private CA.

Creating a Request

data CreatePermission

See: newCreatePermission smart constructor.

Constructors

CreatePermission'

Fields

sourceAccount :: Maybe Text

The ID of the calling account.

certificateAuthorityArn :: Text

The Amazon Resource Name (ARN) of the CA that grants the permissions. You can find the ARN by calling the ListCertificateAuthorities action. This must have the following form:

arn:aws:acm-pca:region:account:certificate-authority/12345678-1234-1234-1234-123456789012 .

principal :: Text

The AWS service or identity that receives the permission. At this time, the only valid principal is acm.amazonaws.com.

actions :: NonEmpty ActionType

The actions that the specified AWS service principal can use. These include IssueCertificate, GetCertificate, and ListPermissions.

newCreatePermission

Arguments

:: Text
$sel:certificateAuthorityArn:CreatePermission'

-> Text
$sel:principal:CreatePermission'

-> NonEmpty ActionType
$sel:actions:CreatePermission'

-> CreatePermission

Create a value of CreatePermission with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:sourceAccount:CreatePermission', createPermission_sourceAccount - The ID of the calling account.

$sel:certificateAuthorityArn:CreatePermission', createPermission_certificateAuthorityArn - The Amazon Resource Name (ARN) of the CA that grants the permissions. You can find the ARN by calling the ListCertificateAuthorities action. This must have the following form:

arn:aws:acm-pca:region:account:certificate-authority/12345678-1234-1234-1234-123456789012 .

$sel:principal:CreatePermission', createPermission_principal - The AWS service or identity that receives the permission. At this time, the only valid principal is acm.amazonaws.com.

$sel:actions:CreatePermission', createPermission_actions - The actions that the specified AWS service principal can use. These include IssueCertificate, GetCertificate, and ListPermissions.

Request Lenses

createPermission_sourceAccount :: Lens' CreatePermission (Maybe Text)

The ID of the calling account.

createPermission_certificateAuthorityArn :: Lens' CreatePermission Text

The Amazon Resource Name (ARN) of the CA that grants the permissions. You can find the ARN by calling the ListCertificateAuthorities action. This must have the following form:

arn:aws:acm-pca:region:account:certificate-authority/12345678-1234-1234-1234-123456789012 .

createPermission_principal :: Lens' CreatePermission Text

The AWS service or identity that receives the permission. At this time, the only valid principal is acm.amazonaws.com.

createPermission_actions :: Lens' CreatePermission (NonEmpty ActionType)

The actions that the specified AWS service principal can use. These include IssueCertificate, GetCertificate, and ListPermissions.

Destructuring the Response

data CreatePermissionResponse

See: newCreatePermissionResponse smart constructor.

Constructors

CreatePermissionResponse'

newCreatePermissionResponse :: CreatePermissionResponse

Create a value of CreatePermissionResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.