Amazonka.CertificateManagerPCA.DescribeCertificateAuthorityAuditReport

Description

Lists information about a specific audit report created by calling the CreateCertificateAuthorityAuditReport action. Audit information is created every time the certificate authority (CA) private key is used. The private key is used when you call the IssueCertificate action or the RevokeCertificate action.

Creating a Request

data DescribeCertificateAuthorityAuditReport

See: newDescribeCertificateAuthorityAuditReport smart constructor.

Constructors

DescribeCertificateAuthorityAuditReport'

Fields

certificateAuthorityArn :: Text

The Amazon Resource Name (ARN) of the private CA. This must be of the form:

arn:aws:acm-pca:region:account:certificate-authority/12345678-1234-1234-1234-123456789012 .

auditReportId :: Text

The report ID returned by calling the CreateCertificateAuthorityAuditReport action.

newDescribeCertificateAuthorityAuditReport

Arguments

:: Text
$sel:certificateAuthorityArn:DescribeCertificateAuthorityAuditReport'

-> Text
$sel:auditReportId:DescribeCertificateAuthorityAuditReport'

-> DescribeCertificateAuthorityAuditReport

Create a value of DescribeCertificateAuthorityAuditReport with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:certificateAuthorityArn:DescribeCertificateAuthorityAuditReport', describeCertificateAuthorityAuditReport_certificateAuthorityArn - The Amazon Resource Name (ARN) of the private CA. This must be of the form:

arn:aws:acm-pca:region:account:certificate-authority/12345678-1234-1234-1234-123456789012 .

$sel:auditReportId:DescribeCertificateAuthorityAuditReport', describeCertificateAuthorityAuditReport_auditReportId - The report ID returned by calling the CreateCertificateAuthorityAuditReport action.

Request Lenses

describeCertificateAuthorityAuditReport_certificateAuthorityArn :: Lens' DescribeCertificateAuthorityAuditReport Text

The Amazon Resource Name (ARN) of the private CA. This must be of the form:

arn:aws:acm-pca:region:account:certificate-authority/12345678-1234-1234-1234-123456789012 .

describeCertificateAuthorityAuditReport_auditReportId :: Lens' DescribeCertificateAuthorityAuditReport Text

The report ID returned by calling the CreateCertificateAuthorityAuditReport action.

Destructuring the Response

data DescribeCertificateAuthorityAuditReportResponse

See: newDescribeCertificateAuthorityAuditReportResponse smart constructor.

Constructors

DescribeCertificateAuthorityAuditReportResponse'

Fields

s3Key :: Maybe Text

S3 key that uniquely identifies the report file in your S3 bucket.

createdAt :: Maybe POSIX

The date and time at which the report was created.

auditReportStatus :: Maybe AuditReportStatus

Specifies whether report creation is in progress, has succeeded, or has failed.

s3BucketName :: Maybe Text

Name of the S3 bucket that contains the report.

httpStatus :: Int

The response's http status code.

newDescribeCertificateAuthorityAuditReportResponse

Arguments

:: Int
$sel:httpStatus:DescribeCertificateAuthorityAuditReportResponse'

-> DescribeCertificateAuthorityAuditReportResponse

Create a value of DescribeCertificateAuthorityAuditReportResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:s3Key:DescribeCertificateAuthorityAuditReportResponse', describeCertificateAuthorityAuditReportResponse_s3Key - S3 key that uniquely identifies the report file in your S3 bucket.

$sel:createdAt:DescribeCertificateAuthorityAuditReportResponse', describeCertificateAuthorityAuditReportResponse_createdAt - The date and time at which the report was created.

$sel:auditReportStatus:DescribeCertificateAuthorityAuditReportResponse', describeCertificateAuthorityAuditReportResponse_auditReportStatus - Specifies whether report creation is in progress, has succeeded, or has failed.

$sel:s3BucketName:DescribeCertificateAuthorityAuditReportResponse', describeCertificateAuthorityAuditReportResponse_s3BucketName - Name of the S3 bucket that contains the report.

$sel:httpStatus:DescribeCertificateAuthorityAuditReportResponse', describeCertificateAuthorityAuditReportResponse_httpStatus - The response's http status code.

Response Lenses

describeCertificateAuthorityAuditReportResponse_s3Key :: Lens' DescribeCertificateAuthorityAuditReportResponse (Maybe Text)

S3 key that uniquely identifies the report file in your S3 bucket.

describeCertificateAuthorityAuditReportResponse_createdAt :: Lens' DescribeCertificateAuthorityAuditReportResponse (Maybe UTCTime)

The date and time at which the report was created.

describeCertificateAuthorityAuditReportResponse_auditReportStatus :: Lens' DescribeCertificateAuthorityAuditReportResponse (Maybe AuditReportStatus)

Specifies whether report creation is in progress, has succeeded, or has failed.

describeCertificateAuthorityAuditReportResponse_s3BucketName :: Lens' DescribeCertificateAuthorityAuditReportResponse (Maybe Text)

Name of the S3 bucket that contains the report.

describeCertificateAuthorityAuditReportResponse_httpStatus :: Lens' DescribeCertificateAuthorityAuditReportResponse Int

The response's http status code.