Amazonka.CertificateManagerPCA.GetCertificateAuthorityCsr

Description

Retrieves the certificate signing request (CSR) for your private certificate authority (CA). The CSR is created when you call the CreateCertificateAuthority action. Sign the CSR with your ACM Private CA-hosted or on-premises root or subordinate CA. Then import the signed certificate back into ACM Private CA by calling the ImportCertificateAuthorityCertificate action. The CSR is returned as a base64 PEM-encoded string.

Creating a Request

data GetCertificateAuthorityCsr

See: newGetCertificateAuthorityCsr smart constructor.

Constructors

GetCertificateAuthorityCsr'

Fields

certificateAuthorityArn :: Text

The Amazon Resource Name (ARN) that was returned when you called the CreateCertificateAuthority action. This must be of the form:

arn:aws:acm-pca:region:account:certificate-authority/12345678-1234-1234-1234-123456789012

newGetCertificateAuthorityCsr

Arguments

:: Text
$sel:certificateAuthorityArn:GetCertificateAuthorityCsr'

-> GetCertificateAuthorityCsr

Create a value of GetCertificateAuthorityCsr with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:certificateAuthorityArn:GetCertificateAuthorityCsr', getCertificateAuthorityCsr_certificateAuthorityArn - The Amazon Resource Name (ARN) that was returned when you called the CreateCertificateAuthority action. This must be of the form:

arn:aws:acm-pca:region:account:certificate-authority/12345678-1234-1234-1234-123456789012

Request Lenses

getCertificateAuthorityCsr_certificateAuthorityArn :: Lens' GetCertificateAuthorityCsr Text

The Amazon Resource Name (ARN) that was returned when you called the CreateCertificateAuthority action. This must be of the form:

arn:aws:acm-pca:region:account:certificate-authority/12345678-1234-1234-1234-123456789012

Destructuring the Response

data GetCertificateAuthorityCsrResponse

See: newGetCertificateAuthorityCsrResponse smart constructor.

Constructors

GetCertificateAuthorityCsrResponse'

Fields

csr :: Maybe Text

The base64 PEM-encoded certificate signing request (CSR) for your private CA certificate.

httpStatus :: Int

The response's http status code.

newGetCertificateAuthorityCsrResponse

Arguments

:: Int
$sel:httpStatus:GetCertificateAuthorityCsrResponse'

-> GetCertificateAuthorityCsrResponse

Create a value of GetCertificateAuthorityCsrResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:csr:GetCertificateAuthorityCsrResponse', getCertificateAuthorityCsrResponse_csr - The base64 PEM-encoded certificate signing request (CSR) for your private CA certificate.

$sel:httpStatus:GetCertificateAuthorityCsrResponse', getCertificateAuthorityCsrResponse_httpStatus - The response's http status code.

Response Lenses

getCertificateAuthorityCsrResponse_csr :: Lens' GetCertificateAuthorityCsrResponse (Maybe Text)

The base64 PEM-encoded certificate signing request (CSR) for your private CA certificate.

getCertificateAuthorityCsrResponse_httpStatus :: Lens' GetCertificateAuthorityCsrResponse Int

The response's http status code.