Amazonka.CertificateManagerPCA.Types.EdiPartyName

Description

Documentation

data EdiPartyName

Describes an Electronic Data Interchange (EDI) entity as described in as defined in Subject Alternative Name in RFC 5280.

See: newEdiPartyName smart constructor.

Constructors

EdiPartyName'

Fields

nameAssigner :: Maybe Text

Specifies the name assigner.

partyName :: Text

Specifies the party name.

newEdiPartyName

Arguments

:: Text
$sel:partyName:EdiPartyName'

-> EdiPartyName

Create a value of EdiPartyName with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:nameAssigner:EdiPartyName', ediPartyName_nameAssigner - Specifies the name assigner.

$sel:partyName:EdiPartyName', ediPartyName_partyName - Specifies the party name.

ediPartyName_nameAssigner :: Lens' EdiPartyName (Maybe Text)

Specifies the name assigner.

ediPartyName_partyName :: Lens' EdiPartyName Text

Specifies the party name.