Amazonka.CertificateManagerPCA.Types.GeneralName

Description

Documentation

data GeneralName

Describes an ASN.1 X.400 GeneralName as defined in RFC 5280. Only one of the following naming options should be provided. Providing more than one option results in an InvalidArgsException error.

See: newGeneralName smart constructor.

Constructors

GeneralName'

Fields

ipAddress :: Maybe Text

Represents GeneralName as an IPv4 or IPv6 address.

uniformResourceIdentifier :: Maybe Text

Represents GeneralName as a URI.

registeredId :: Maybe Text

Represents GeneralName as an object identifier (OID).

ediPartyName :: Maybe EdiPartyName

Represents GeneralName as an EdiPartyName object.

rfc822Name :: Maybe Text

Represents GeneralName as an RFC 822 email address.

otherName :: Maybe OtherName

Represents GeneralName using an OtherName object.

dnsName :: Maybe Text

Represents GeneralName as a DNS name.

directoryName :: Maybe ASN1Subject

newGeneralName :: GeneralName

Create a value of GeneralName with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:ipAddress:GeneralName', generalName_ipAddress - Represents GeneralName as an IPv4 or IPv6 address.

$sel:uniformResourceIdentifier:GeneralName', generalName_uniformResourceIdentifier - Represents GeneralName as a URI.

$sel:registeredId:GeneralName', generalName_registeredId - Represents GeneralName as an object identifier (OID).

$sel:ediPartyName:GeneralName', generalName_ediPartyName - Represents GeneralName as an EdiPartyName object.

$sel:rfc822Name:GeneralName', generalName_rfc822Name - Represents GeneralName as an RFC 822 email address.

$sel:otherName:GeneralName', generalName_otherName - Represents GeneralName using an OtherName object.

$sel:dnsName:GeneralName', generalName_dnsName - Represents GeneralName as a DNS name.

$sel:directoryName:GeneralName', generalName_directoryName - Undocumented member.

generalName_ipAddress :: Lens' GeneralName (Maybe Text)

Represents GeneralName as an IPv4 or IPv6 address.

generalName_uniformResourceIdentifier :: Lens' GeneralName (Maybe Text)

Represents GeneralName as a URI.

generalName_registeredId :: Lens' GeneralName (Maybe Text)

Represents GeneralName as an object identifier (OID).

generalName_ediPartyName :: Lens' GeneralName (Maybe EdiPartyName)

Represents GeneralName as an EdiPartyName object.

generalName_rfc822Name :: Lens' GeneralName (Maybe Text)

Represents GeneralName as an RFC 822 email address.

generalName_otherName :: Lens' GeneralName (Maybe OtherName)

Represents GeneralName using an OtherName object.

generalName_dnsName :: Lens' GeneralName (Maybe Text)

Represents GeneralName as a DNS name.

generalName_directoryName :: Lens' GeneralName (Maybe ASN1Subject)

Undocumented member.