Amazonka.CertificateManagerPCA.Types.PolicyInformation

Description

Documentation

data PolicyInformation

Defines the X.509 CertificatePolicies extension.

See: newPolicyInformation smart constructor.

Constructors

PolicyInformation'

Fields

policyQualifiers :: Maybe (NonEmpty PolicyQualifierInfo)

Modifies the given CertPolicyId with a qualifier. ACM Private CA supports the certification practice statement (CPS) qualifier.

certPolicyId :: Text

Specifies the object identifier (OID) of the certificate policy under which the certificate was issued. For more information, see NIST's definition of Object Identifier (OID).

newPolicyInformation

Arguments

:: Text
$sel:certPolicyId:PolicyInformation'

-> PolicyInformation

Create a value of PolicyInformation with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:policyQualifiers:PolicyInformation', policyInformation_policyQualifiers - Modifies the given CertPolicyId with a qualifier. ACM Private CA supports the certification practice statement (CPS) qualifier.

$sel:certPolicyId:PolicyInformation', policyInformation_certPolicyId - Specifies the object identifier (OID) of the certificate policy under which the certificate was issued. For more information, see NIST's definition of Object Identifier (OID).

policyInformation_policyQualifiers :: Lens' PolicyInformation (Maybe (NonEmpty PolicyQualifierInfo))

Modifies the given CertPolicyId with a qualifier. ACM Private CA supports the certification practice statement (CPS) qualifier.

policyInformation_certPolicyId :: Lens' PolicyInformation Text

Specifies the object identifier (OID) of the certificate policy under which the certificate was issued. For more information, see NIST's definition of Object Identifier (OID).