Amazonka.CertificateManagerPCA.UntagCertificateAuthority

Description

Remove one or more tags from your private CA. A tag consists of a key-value pair. If you do not specify the value portion of the tag when calling this action, the tag will be removed regardless of value. If you specify a value, the tag is removed only if it is associated with the specified value. To add tags to a private CA, use the TagCertificateAuthority. Call the ListTags action to see what tags are associated with your CA.

Creating a Request

data UntagCertificateAuthority

See: newUntagCertificateAuthority smart constructor.

Constructors

UntagCertificateAuthority'

Fields

certificateAuthorityArn :: Text

The Amazon Resource Name (ARN) that was returned when you called CreateCertificateAuthority. This must be of the form:

arn:aws:acm-pca:region:account:certificate-authority/12345678-1234-1234-1234-123456789012

tags :: NonEmpty Tag

List of tags to be removed from the CA.

newUntagCertificateAuthority

Arguments

:: Text
$sel:certificateAuthorityArn:UntagCertificateAuthority'

-> NonEmpty Tag
$sel:tags:UntagCertificateAuthority'

-> UntagCertificateAuthority

Create a value of UntagCertificateAuthority with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:certificateAuthorityArn:UntagCertificateAuthority', untagCertificateAuthority_certificateAuthorityArn - The Amazon Resource Name (ARN) that was returned when you called CreateCertificateAuthority. This must be of the form:

arn:aws:acm-pca:region:account:certificate-authority/12345678-1234-1234-1234-123456789012

$sel:tags:UntagCertificateAuthority', untagCertificateAuthority_tags - List of tags to be removed from the CA.

Request Lenses

untagCertificateAuthority_certificateAuthorityArn :: Lens' UntagCertificateAuthority Text

The Amazon Resource Name (ARN) that was returned when you called CreateCertificateAuthority. This must be of the form:

arn:aws:acm-pca:region:account:certificate-authority/12345678-1234-1234-1234-123456789012

untagCertificateAuthority_tags :: Lens' UntagCertificateAuthority (NonEmpty Tag)

List of tags to be removed from the CA.

Destructuring the Response

data UntagCertificateAuthorityResponse

See: newUntagCertificateAuthorityResponse smart constructor.

Constructors

UntagCertificateAuthorityResponse'

newUntagCertificateAuthorityResponse :: UntagCertificateAuthorityResponse

Create a value of UntagCertificateAuthorityResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.