Amazonka.CertificateManagerPCA.UpdateCertificateAuthority

Description

Updates the status or configuration of a private certificate authority (CA). Your private CA must be in the ACTIVE or DISABLED state before you can update it. You can disable a private CA that is in the ACTIVE state or make a CA that is in the DISABLED state active again.

Both PCA and the IAM principal must have permission to write to the S3 bucket that you specify. If the IAM principal making the call does not have permission to write to the bucket, then an exception is thrown. For more information, see Configure Access to ACM Private CA.

Creating a Request

data UpdateCertificateAuthority

See: newUpdateCertificateAuthority smart constructor.

Constructors

UpdateCertificateAuthority'

Fields

status :: Maybe CertificateAuthorityStatus

Status of your private CA.

revocationConfiguration :: Maybe RevocationConfiguration

Contains information to enable Online Certificate Status Protocol (OCSP) support, to enable a certificate revocation list (CRL), to enable both, or to enable neither. If this parameter is not supplied, existing capibilites remain unchanged. For more information, see the OcspConfiguration and CrlConfiguration types.

certificateAuthorityArn :: Text

Amazon Resource Name (ARN) of the private CA that issued the certificate to be revoked. This must be of the form:

arn:aws:acm-pca:region:account:certificate-authority/12345678-1234-1234-1234-123456789012

newUpdateCertificateAuthority

Arguments

:: Text
$sel:certificateAuthorityArn:UpdateCertificateAuthority'

-> UpdateCertificateAuthority

Create a value of UpdateCertificateAuthority with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:status:UpdateCertificateAuthority', updateCertificateAuthority_status - Status of your private CA.

$sel:revocationConfiguration:UpdateCertificateAuthority', updateCertificateAuthority_revocationConfiguration - Contains information to enable Online Certificate Status Protocol (OCSP) support, to enable a certificate revocation list (CRL), to enable both, or to enable neither. If this parameter is not supplied, existing capibilites remain unchanged. For more information, see the OcspConfiguration and CrlConfiguration types.

$sel:certificateAuthorityArn:UpdateCertificateAuthority', updateCertificateAuthority_certificateAuthorityArn - Amazon Resource Name (ARN) of the private CA that issued the certificate to be revoked. This must be of the form:

arn:aws:acm-pca:region:account:certificate-authority/12345678-1234-1234-1234-123456789012

Request Lenses

updateCertificateAuthority_status :: Lens' UpdateCertificateAuthority (Maybe CertificateAuthorityStatus)

Status of your private CA.

updateCertificateAuthority_revocationConfiguration :: Lens' UpdateCertificateAuthority (Maybe RevocationConfiguration)

Contains information to enable Online Certificate Status Protocol (OCSP) support, to enable a certificate revocation list (CRL), to enable both, or to enable neither. If this parameter is not supplied, existing capibilites remain unchanged. For more information, see the OcspConfiguration and CrlConfiguration types.

updateCertificateAuthority_certificateAuthorityArn :: Lens' UpdateCertificateAuthority Text

Amazon Resource Name (ARN) of the private CA that issued the certificate to be revoked. This must be of the form:

arn:aws:acm-pca:region:account:certificate-authority/12345678-1234-1234-1234-123456789012

Destructuring the Response

data UpdateCertificateAuthorityResponse

See: newUpdateCertificateAuthorityResponse smart constructor.

Constructors

UpdateCertificateAuthorityResponse'

newUpdateCertificateAuthorityResponse :: UpdateCertificateAuthorityResponse

Create a value of UpdateCertificateAuthorityResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.