Amazonka.Chime.AssociatePhoneNumbersWithVoiceConnector

Description

Associates phone numbers with the specified Amazon Chime Voice Connector.

Creating a Request

data AssociatePhoneNumbersWithVoiceConnector

See: newAssociatePhoneNumbersWithVoiceConnector smart constructor.

Constructors

AssociatePhoneNumbersWithVoiceConnector'

Fields

forceAssociate :: Maybe Bool

If true, associates the provided phone numbers with the provided Amazon Chime Voice Connector and removes any previously existing associations. If false, does not associate any phone numbers that have previously existing associations.

voiceConnectorId :: Text

The Amazon Chime Voice Connector ID.

e164PhoneNumbers :: [Sensitive Text]

List of phone numbers, in E.164 format.

newAssociatePhoneNumbersWithVoiceConnector

Arguments

:: Text
$sel:voiceConnectorId:AssociatePhoneNumbersWithVoiceConnector'

-> AssociatePhoneNumbersWithVoiceConnector

Create a value of AssociatePhoneNumbersWithVoiceConnector with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:forceAssociate:AssociatePhoneNumbersWithVoiceConnector', associatePhoneNumbersWithVoiceConnector_forceAssociate - If true, associates the provided phone numbers with the provided Amazon Chime Voice Connector and removes any previously existing associations. If false, does not associate any phone numbers that have previously existing associations.

$sel:voiceConnectorId:AssociatePhoneNumbersWithVoiceConnector', associatePhoneNumbersWithVoiceConnector_voiceConnectorId - The Amazon Chime Voice Connector ID.

$sel:e164PhoneNumbers:AssociatePhoneNumbersWithVoiceConnector', associatePhoneNumbersWithVoiceConnector_e164PhoneNumbers - List of phone numbers, in E.164 format.

Request Lenses

associatePhoneNumbersWithVoiceConnector_forceAssociate :: Lens' AssociatePhoneNumbersWithVoiceConnector (Maybe Bool)

If true, associates the provided phone numbers with the provided Amazon Chime Voice Connector and removes any previously existing associations. If false, does not associate any phone numbers that have previously existing associations.

associatePhoneNumbersWithVoiceConnector_voiceConnectorId :: Lens' AssociatePhoneNumbersWithVoiceConnector Text

The Amazon Chime Voice Connector ID.

associatePhoneNumbersWithVoiceConnector_e164PhoneNumbers :: Lens' AssociatePhoneNumbersWithVoiceConnector [Text]

List of phone numbers, in E.164 format.

Destructuring the Response

data AssociatePhoneNumbersWithVoiceConnectorResponse

See: newAssociatePhoneNumbersWithVoiceConnectorResponse smart constructor.

Constructors

AssociatePhoneNumbersWithVoiceConnectorResponse'

Fields

phoneNumberErrors :: Maybe [PhoneNumberError]

If the action fails for one or more of the phone numbers in the request, a list of the phone numbers is returned, along with error codes and error messages.

httpStatus :: Int

The response's http status code.

newAssociatePhoneNumbersWithVoiceConnectorResponse

Arguments

:: Int
$sel:httpStatus:AssociatePhoneNumbersWithVoiceConnectorResponse'

-> AssociatePhoneNumbersWithVoiceConnectorResponse

Create a value of AssociatePhoneNumbersWithVoiceConnectorResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:phoneNumberErrors:AssociatePhoneNumbersWithVoiceConnectorResponse', associatePhoneNumbersWithVoiceConnectorResponse_phoneNumberErrors - If the action fails for one or more of the phone numbers in the request, a list of the phone numbers is returned, along with error codes and error messages.

$sel:httpStatus:AssociatePhoneNumbersWithVoiceConnectorResponse', associatePhoneNumbersWithVoiceConnectorResponse_httpStatus - The response's http status code.

Response Lenses

associatePhoneNumbersWithVoiceConnectorResponse_phoneNumberErrors :: Lens' AssociatePhoneNumbersWithVoiceConnectorResponse (Maybe [PhoneNumberError])

If the action fails for one or more of the phone numbers in the request, a list of the phone numbers is returned, along with error codes and error messages.

associatePhoneNumbersWithVoiceConnectorResponse_httpStatus :: Lens' AssociatePhoneNumbersWithVoiceConnectorResponse Int

The response's http status code.