Amazonka.Chime.AssociateSigninDelegateGroupsWithAccount

Description

Associates the specified sign-in delegate groups with the specified Amazon Chime account.

Creating a Request

data AssociateSigninDelegateGroupsWithAccount

See: newAssociateSigninDelegateGroupsWithAccount smart constructor.

Constructors

AssociateSigninDelegateGroupsWithAccount'

Fields

accountId :: Text

The Amazon Chime account ID.

signinDelegateGroups :: [SigninDelegateGroup]

The sign-in delegate groups.

newAssociateSigninDelegateGroupsWithAccount

Arguments

:: Text
$sel:accountId:AssociateSigninDelegateGroupsWithAccount'

-> AssociateSigninDelegateGroupsWithAccount

Create a value of AssociateSigninDelegateGroupsWithAccount with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:accountId:AssociateSigninDelegateGroupsWithAccount', associateSigninDelegateGroupsWithAccount_accountId - The Amazon Chime account ID.

$sel:signinDelegateGroups:AssociateSigninDelegateGroupsWithAccount', associateSigninDelegateGroupsWithAccount_signinDelegateGroups - The sign-in delegate groups.

Request Lenses

associateSigninDelegateGroupsWithAccount_accountId :: Lens' AssociateSigninDelegateGroupsWithAccount Text

The Amazon Chime account ID.

associateSigninDelegateGroupsWithAccount_signinDelegateGroups :: Lens' AssociateSigninDelegateGroupsWithAccount [SigninDelegateGroup]

The sign-in delegate groups.

Destructuring the Response

data AssociateSigninDelegateGroupsWithAccountResponse

See: newAssociateSigninDelegateGroupsWithAccountResponse smart constructor.

Constructors

AssociateSigninDelegateGroupsWithAccountResponse'

Fields

httpStatus :: Int

The response's http status code.

newAssociateSigninDelegateGroupsWithAccountResponse

Arguments

:: Int
$sel:httpStatus:AssociateSigninDelegateGroupsWithAccountResponse'

-> AssociateSigninDelegateGroupsWithAccountResponse

Create a value of AssociateSigninDelegateGroupsWithAccountResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:httpStatus:AssociateSigninDelegateGroupsWithAccountResponse', associateSigninDelegateGroupsWithAccountResponse_httpStatus - The response's http status code.

Response Lenses

associateSigninDelegateGroupsWithAccountResponse_httpStatus :: Lens' AssociateSigninDelegateGroupsWithAccountResponse Int

The response's http status code.