Amazonka.Chime.BatchUpdatePhoneNumber

Description

Updates phone number product types or calling names. You can update one attribute at a time for each UpdatePhoneNumberRequestItem. For example, you can update the product type or the calling name.

For toll-free numbers, you cannot use the Amazon Chime Business Calling product type. For numbers outside the U.S., you must use the Amazon Chime SIP Media Application Dial-In product type.

Updates to outbound calling names can take up to 72 hours to complete. Pending updates to outbound calling names must be complete before you can request another update.

Creating a Request

data BatchUpdatePhoneNumber

See: newBatchUpdatePhoneNumber smart constructor.

Constructors

BatchUpdatePhoneNumber'

Fields

updatePhoneNumberRequestItems :: [UpdatePhoneNumberRequestItem]

The request containing the phone number IDs and product types or calling names to update.

newBatchUpdatePhoneNumber :: BatchUpdatePhoneNumber

Create a value of BatchUpdatePhoneNumber with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:updatePhoneNumberRequestItems:BatchUpdatePhoneNumber', batchUpdatePhoneNumber_updatePhoneNumberRequestItems - The request containing the phone number IDs and product types or calling names to update.

Request Lenses

batchUpdatePhoneNumber_updatePhoneNumberRequestItems :: Lens' BatchUpdatePhoneNumber [UpdatePhoneNumberRequestItem]

The request containing the phone number IDs and product types or calling names to update.

Destructuring the Response

data BatchUpdatePhoneNumberResponse

See: newBatchUpdatePhoneNumberResponse smart constructor.

Constructors

BatchUpdatePhoneNumberResponse'

Fields

phoneNumberErrors :: Maybe [PhoneNumberError]

If the action fails for one or more of the phone numbers in the request, a list of the phone numbers is returned, along with error codes and error messages.

httpStatus :: Int

The response's http status code.

newBatchUpdatePhoneNumberResponse

Arguments

:: Int
$sel:httpStatus:BatchUpdatePhoneNumberResponse'

-> BatchUpdatePhoneNumberResponse

Create a value of BatchUpdatePhoneNumberResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:phoneNumberErrors:BatchUpdatePhoneNumberResponse', batchUpdatePhoneNumberResponse_phoneNumberErrors - If the action fails for one or more of the phone numbers in the request, a list of the phone numbers is returned, along with error codes and error messages.

$sel:httpStatus:BatchUpdatePhoneNumberResponse', batchUpdatePhoneNumberResponse_httpStatus - The response's http status code.

Response Lenses

batchUpdatePhoneNumberResponse_phoneNumberErrors :: Lens' BatchUpdatePhoneNumberResponse (Maybe [PhoneNumberError])

If the action fails for one or more of the phone numbers in the request, a list of the phone numbers is returned, along with error codes and error messages.

batchUpdatePhoneNumberResponse_httpStatus :: Lens' BatchUpdatePhoneNumberResponse Int

The response's http status code.