Amazonka.Chime.CreateAppInstanceAdmin

Description

Promotes an AppInstanceUser to an AppInstanceAdmin. The promoted user can perform the following actions.

ChannelModerator actions across all channels in the AppInstance.

DeleteChannelMessage actions.

Only an AppInstanceUser can be promoted to an AppInstanceAdmin role.

Creating a Request

data CreateAppInstanceAdmin

See: newCreateAppInstanceAdmin smart constructor.

Constructors

CreateAppInstanceAdmin'

Fields

appInstanceAdminArn :: Text

The ARN of the administrator of the current AppInstance.

appInstanceArn :: Text

The ARN of the AppInstance.

newCreateAppInstanceAdmin

Arguments

:: Text
$sel:appInstanceAdminArn:CreateAppInstanceAdmin'

-> Text
$sel:appInstanceArn:CreateAppInstanceAdmin'

-> CreateAppInstanceAdmin

Create a value of CreateAppInstanceAdmin with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:appInstanceAdminArn:CreateAppInstanceAdmin', createAppInstanceAdmin_appInstanceAdminArn - The ARN of the administrator of the current AppInstance.

$sel:appInstanceArn:CreateAppInstanceAdmin', createAppInstanceAdmin_appInstanceArn - The ARN of the AppInstance.

Request Lenses

createAppInstanceAdmin_appInstanceAdminArn :: Lens' CreateAppInstanceAdmin Text

The ARN of the administrator of the current AppInstance.

createAppInstanceAdmin_appInstanceArn :: Lens' CreateAppInstanceAdmin Text

The ARN of the AppInstance.

Destructuring the Response

data CreateAppInstanceAdminResponse

See: newCreateAppInstanceAdminResponse smart constructor.

Constructors

CreateAppInstanceAdminResponse'

Fields

appInstanceAdmin :: Maybe Identity

The name and ARN of the admin for the AppInstance.

appInstanceArn :: Maybe Text

The ARN of the of the admin for the AppInstance.

httpStatus :: Int

The response's http status code.

newCreateAppInstanceAdminResponse

Arguments

:: Int
$sel:httpStatus:CreateAppInstanceAdminResponse'

-> CreateAppInstanceAdminResponse

Create a value of CreateAppInstanceAdminResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:appInstanceAdmin:CreateAppInstanceAdminResponse', createAppInstanceAdminResponse_appInstanceAdmin - The name and ARN of the admin for the AppInstance.

$sel:appInstanceArn:CreateAppInstanceAdmin', createAppInstanceAdminResponse_appInstanceArn - The ARN of the of the admin for the AppInstance.

$sel:httpStatus:CreateAppInstanceAdminResponse', createAppInstanceAdminResponse_httpStatus - The response's http status code.

Response Lenses

createAppInstanceAdminResponse_appInstanceAdmin :: Lens' CreateAppInstanceAdminResponse (Maybe Identity)

The name and ARN of the admin for the AppInstance.

createAppInstanceAdminResponse_appInstanceArn :: Lens' CreateAppInstanceAdminResponse (Maybe Text)

The ARN of the of the admin for the AppInstance.

createAppInstanceAdminResponse_httpStatus :: Lens' CreateAppInstanceAdminResponse Int

The response's http status code.