Amazonka.Chime.CreateAttendee

Description

Creates a new attendee for an active Amazon Chime SDK meeting. For more information about the Amazon Chime SDK, see Using the Amazon Chime SDK in the Amazon Chime Developer Guide.

Creating a Request

data CreateAttendee

See: newCreateAttendee smart constructor.

Constructors

CreateAttendee'

Fields

tags :: Maybe (NonEmpty Tag)

The tag key-value pairs.

meetingId :: Text

The Amazon Chime SDK meeting ID.

externalUserId :: Sensitive Text

The Amazon Chime SDK external user ID. An idempotency token. Links the attendee to an identity managed by a builder application.

newCreateAttendee

Arguments

:: Text
$sel:meetingId:CreateAttendee'

-> Text
$sel:externalUserId:CreateAttendee'

-> CreateAttendee

Create a value of CreateAttendee with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:tags:CreateAttendee', createAttendee_tags - The tag key-value pairs.

$sel:meetingId:CreateAttendee', createAttendee_meetingId - The Amazon Chime SDK meeting ID.

$sel:externalUserId:CreateAttendee', createAttendee_externalUserId - The Amazon Chime SDK external user ID. An idempotency token. Links the attendee to an identity managed by a builder application.

Request Lenses

createAttendee_tags :: Lens' CreateAttendee (Maybe (NonEmpty Tag))

The tag key-value pairs.

createAttendee_meetingId :: Lens' CreateAttendee Text

The Amazon Chime SDK meeting ID.

createAttendee_externalUserId :: Lens' CreateAttendee Text

The Amazon Chime SDK external user ID. An idempotency token. Links the attendee to an identity managed by a builder application.

Destructuring the Response

data CreateAttendeeResponse

See: newCreateAttendeeResponse smart constructor.

Constructors

CreateAttendeeResponse'

Fields

attendee :: Maybe Attendee

The attendee information, including attendee ID and join token.

httpStatus :: Int

The response's http status code.

newCreateAttendeeResponse

Arguments

:: Int
$sel:httpStatus:CreateAttendeeResponse'

-> CreateAttendeeResponse

Create a value of CreateAttendeeResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:attendee:CreateAttendeeResponse', createAttendeeResponse_attendee - The attendee information, including attendee ID and join token.

$sel:httpStatus:CreateAttendeeResponse', createAttendeeResponse_httpStatus - The response's http status code.

Response Lenses

createAttendeeResponse_attendee :: Lens' CreateAttendeeResponse (Maybe Attendee)

The attendee information, including attendee ID and join token.

createAttendeeResponse_httpStatus :: Lens' CreateAttendeeResponse Int

The response's http status code.