Amazonka.Chime.CreateBot

Description

Creates a bot for an Amazon Chime Enterprise account.

Creating a Request

data CreateBot

See: newCreateBot smart constructor.

Constructors

CreateBot'

Fields

domain :: Maybe Text

The domain of the Amazon Chime Enterprise account.

displayName :: Sensitive Text

The bot display name.

accountId :: Text

The Amazon Chime account ID.

newCreateBot

Arguments

:: Text
$sel:displayName:CreateBot'

-> Text
$sel:accountId:CreateBot'

-> CreateBot

Create a value of CreateBot with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:domain:CreateBot', createBot_domain - The domain of the Amazon Chime Enterprise account.

$sel:displayName:CreateBot', createBot_displayName - The bot display name.

$sel:accountId:CreateBot', createBot_accountId - The Amazon Chime account ID.

Request Lenses

createBot_domain :: Lens' CreateBot (Maybe Text)

The domain of the Amazon Chime Enterprise account.

createBot_displayName :: Lens' CreateBot Text

The bot display name.

createBot_accountId :: Lens' CreateBot Text

The Amazon Chime account ID.

Destructuring the Response

data CreateBotResponse

See: newCreateBotResponse smart constructor.

Constructors

CreateBotResponse'

Fields

bot :: Maybe Bot

The bot details.

httpStatus :: Int

The response's http status code.

newCreateBotResponse

Arguments

:: Int
$sel:httpStatus:CreateBotResponse'

-> CreateBotResponse

Create a value of CreateBotResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:bot:CreateBotResponse', createBotResponse_bot - The bot details.

$sel:httpStatus:CreateBotResponse', createBotResponse_httpStatus - The response's http status code.

Response Lenses

createBotResponse_bot :: Lens' CreateBotResponse (Maybe Bot)

The bot details.

createBotResponse_httpStatus :: Lens' CreateBotResponse Int

The response's http status code.