Amazonka.Chime.CreateMediaCapturePipeline

Description

Creates a media capture pipeline.

Creating a Request

data CreateMediaCapturePipeline

See: newCreateMediaCapturePipeline smart constructor.

Constructors

CreateMediaCapturePipeline'

Fields

chimeSdkMeetingConfiguration :: Maybe ChimeSdkMeetingConfiguration

The configuration for a specified media capture pipeline. SourceType must be ChimeSdkMeeting.

clientRequestToken :: Maybe (Sensitive Text)

The token assigned to the client making the pipeline request.

sourceType :: MediaPipelineSourceType

Source type from which the media artifacts will be captured. A Chime SDK Meeting is the only supported source.

sourceArn :: Sensitive Text

ARN of the source from which the media artifacts are captured.

sinkType :: MediaPipelineSinkType

Destination type to which the media artifacts are saved. You must use an S3 bucket.

sinkArn :: Sensitive Text

The ARN of the sink type.

newCreateMediaCapturePipeline

Arguments

:: MediaPipelineSourceType
$sel:sourceType:CreateMediaCapturePipeline'

-> Text
$sel:sourceArn:CreateMediaCapturePipeline'

-> MediaPipelineSinkType
$sel:sinkType:CreateMediaCapturePipeline'

-> Text
$sel:sinkArn:CreateMediaCapturePipeline'

-> CreateMediaCapturePipeline

Create a value of CreateMediaCapturePipeline with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:chimeSdkMeetingConfiguration:CreateMediaCapturePipeline', createMediaCapturePipeline_chimeSdkMeetingConfiguration - The configuration for a specified media capture pipeline. SourceType must be ChimeSdkMeeting.

$sel:clientRequestToken:CreateMediaCapturePipeline', createMediaCapturePipeline_clientRequestToken - The token assigned to the client making the pipeline request.

$sel:sourceType:CreateMediaCapturePipeline', createMediaCapturePipeline_sourceType - Source type from which the media artifacts will be captured. A Chime SDK Meeting is the only supported source.

$sel:sourceArn:CreateMediaCapturePipeline', createMediaCapturePipeline_sourceArn - ARN of the source from which the media artifacts are captured.

$sel:sinkType:CreateMediaCapturePipeline', createMediaCapturePipeline_sinkType - Destination type to which the media artifacts are saved. You must use an S3 bucket.

$sel:sinkArn:CreateMediaCapturePipeline', createMediaCapturePipeline_sinkArn - The ARN of the sink type.

Request Lenses

createMediaCapturePipeline_chimeSdkMeetingConfiguration :: Lens' CreateMediaCapturePipeline (Maybe ChimeSdkMeetingConfiguration)

The configuration for a specified media capture pipeline. SourceType must be ChimeSdkMeeting.

createMediaCapturePipeline_clientRequestToken :: Lens' CreateMediaCapturePipeline (Maybe Text)

The token assigned to the client making the pipeline request.

createMediaCapturePipeline_sourceType :: Lens' CreateMediaCapturePipeline MediaPipelineSourceType

Source type from which the media artifacts will be captured. A Chime SDK Meeting is the only supported source.

createMediaCapturePipeline_sourceArn :: Lens' CreateMediaCapturePipeline Text

ARN of the source from which the media artifacts are captured.

createMediaCapturePipeline_sinkType :: Lens' CreateMediaCapturePipeline MediaPipelineSinkType

Destination type to which the media artifacts are saved. You must use an S3 bucket.

createMediaCapturePipeline_sinkArn :: Lens' CreateMediaCapturePipeline Text

The ARN of the sink type.

Destructuring the Response

data CreateMediaCapturePipelineResponse

See: newCreateMediaCapturePipelineResponse smart constructor.

Constructors

CreateMediaCapturePipelineResponse'

Fields

mediaCapturePipeline :: Maybe MediaCapturePipeline

A media capture pipeline object, the ID, source type, source ARN, sink type, and sink ARN of a media capture pipeline object.

httpStatus :: Int

The response's http status code.

newCreateMediaCapturePipelineResponse

Arguments

:: Int
$sel:httpStatus:CreateMediaCapturePipelineResponse'

-> CreateMediaCapturePipelineResponse

Create a value of CreateMediaCapturePipelineResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:mediaCapturePipeline:CreateMediaCapturePipelineResponse', createMediaCapturePipelineResponse_mediaCapturePipeline - A media capture pipeline object, the ID, source type, source ARN, sink type, and sink ARN of a media capture pipeline object.

$sel:httpStatus:CreateMediaCapturePipelineResponse', createMediaCapturePipelineResponse_httpStatus - The response's http status code.

Response Lenses

createMediaCapturePipelineResponse_mediaCapturePipeline :: Lens' CreateMediaCapturePipelineResponse (Maybe MediaCapturePipeline)

A media capture pipeline object, the ID, source type, source ARN, sink type, and sink ARN of a media capture pipeline object.

createMediaCapturePipelineResponse_httpStatus :: Lens' CreateMediaCapturePipelineResponse Int

The response's http status code.