Amazonka.Chime.CreatePhoneNumberOrder

Description

Creates an order for phone numbers to be provisioned. For toll-free numbers, you cannot use the Amazon Chime Business Calling product type. For numbers outside the U.S., you must use the Amazon Chime SIP Media Application Dial-In product type.

Creating a Request

data CreatePhoneNumberOrder

See: newCreatePhoneNumberOrder smart constructor.

Constructors

CreatePhoneNumberOrder'

Fields

productType :: PhoneNumberProductType

The phone number product type.

e164PhoneNumbers :: [Sensitive Text]

List of phone numbers, in E.164 format.

newCreatePhoneNumberOrder

Arguments

:: PhoneNumberProductType
$sel:productType:CreatePhoneNumberOrder'

-> CreatePhoneNumberOrder

Create a value of CreatePhoneNumberOrder with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:productType:CreatePhoneNumberOrder', createPhoneNumberOrder_productType - The phone number product type.

$sel:e164PhoneNumbers:CreatePhoneNumberOrder', createPhoneNumberOrder_e164PhoneNumbers - List of phone numbers, in E.164 format.

Request Lenses

createPhoneNumberOrder_productType :: Lens' CreatePhoneNumberOrder PhoneNumberProductType

The phone number product type.

createPhoneNumberOrder_e164PhoneNumbers :: Lens' CreatePhoneNumberOrder [Text]

List of phone numbers, in E.164 format.

Destructuring the Response

data CreatePhoneNumberOrderResponse

See: newCreatePhoneNumberOrderResponse smart constructor.

Constructors

CreatePhoneNumberOrderResponse'

Fields

phoneNumberOrder :: Maybe PhoneNumberOrder

The phone number order details.

httpStatus :: Int

The response's http status code.

newCreatePhoneNumberOrderResponse

Arguments

:: Int
$sel:httpStatus:CreatePhoneNumberOrderResponse'

-> CreatePhoneNumberOrderResponse

Create a value of CreatePhoneNumberOrderResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:phoneNumberOrder:CreatePhoneNumberOrderResponse', createPhoneNumberOrderResponse_phoneNumberOrder - The phone number order details.

$sel:httpStatus:CreatePhoneNumberOrderResponse', createPhoneNumberOrderResponse_httpStatus - The response's http status code.

Response Lenses

createPhoneNumberOrderResponse_phoneNumberOrder :: Lens' CreatePhoneNumberOrderResponse (Maybe PhoneNumberOrder)

The phone number order details.

createPhoneNumberOrderResponse_httpStatus :: Lens' CreatePhoneNumberOrderResponse Int

The response's http status code.