Amazonka.Chime.CreateRoom

Description

Creates a chat room for the specified Amazon Chime Enterprise account.

Creating a Request

data CreateRoom

See: newCreateRoom smart constructor.

Constructors

CreateRoom'

Fields

clientRequestToken :: Maybe (Sensitive Text)

The idempotency token for the request.

accountId :: Text

The Amazon Chime account ID.

name :: Sensitive Text

The room name.

newCreateRoom

Arguments

:: Text
$sel:accountId:CreateRoom'

-> Text
$sel:name:CreateRoom'

-> CreateRoom

Create a value of CreateRoom with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:clientRequestToken:CreateRoom', createRoom_clientRequestToken - The idempotency token for the request.

$sel:accountId:CreateRoom', createRoom_accountId - The Amazon Chime account ID.

$sel:name:CreateRoom', createRoom_name - The room name.

Request Lenses

createRoom_clientRequestToken :: Lens' CreateRoom (Maybe Text)

The idempotency token for the request.

createRoom_accountId :: Lens' CreateRoom Text

The Amazon Chime account ID.

createRoom_name :: Lens' CreateRoom Text

The room name.

Destructuring the Response

data CreateRoomResponse

See: newCreateRoomResponse smart constructor.

Constructors

CreateRoomResponse'

Fields

room :: Maybe Room

The room details.

httpStatus :: Int

The response's http status code.

newCreateRoomResponse

Arguments

:: Int
$sel:httpStatus:CreateRoomResponse'

-> CreateRoomResponse

Create a value of CreateRoomResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:room:CreateRoomResponse', createRoomResponse_room - The room details.

$sel:httpStatus:CreateRoomResponse', createRoomResponse_httpStatus - The response's http status code.

Response Lenses

createRoomResponse_room :: Lens' CreateRoomResponse (Maybe Room)

The room details.

createRoomResponse_httpStatus :: Lens' CreateRoomResponse Int

The response's http status code.