Amazonka.Chime.CreateRoomMembership

Description

Adds a member to a chat room in an Amazon Chime Enterprise account. A member can be either a user or a bot. The member role designates whether the member is a chat room administrator or a general chat room member.

Creating a Request

data CreateRoomMembership

See: newCreateRoomMembership smart constructor.

Constructors

CreateRoomMembership'

Fields

role' :: Maybe RoomMembershipRole

The role of the member.

accountId :: Text

The Amazon Chime account ID.

roomId :: Text

The room ID.

memberId :: Text

The Amazon Chime member ID (user ID or bot ID).

newCreateRoomMembership

Arguments

:: Text
$sel:accountId:CreateRoomMembership'

-> Text
$sel:roomId:CreateRoomMembership'

-> Text
$sel:memberId:CreateRoomMembership'

-> CreateRoomMembership

Create a value of CreateRoomMembership with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:role':CreateRoomMembership', createRoomMembership_role - The role of the member.

$sel:accountId:CreateRoomMembership', createRoomMembership_accountId - The Amazon Chime account ID.

$sel:roomId:CreateRoomMembership', createRoomMembership_roomId - The room ID.

$sel:memberId:CreateRoomMembership', createRoomMembership_memberId - The Amazon Chime member ID (user ID or bot ID).

Request Lenses

createRoomMembership_role :: Lens' CreateRoomMembership (Maybe RoomMembershipRole)

The role of the member.

createRoomMembership_accountId :: Lens' CreateRoomMembership Text

The Amazon Chime account ID.

createRoomMembership_roomId :: Lens' CreateRoomMembership Text

The room ID.

createRoomMembership_memberId :: Lens' CreateRoomMembership Text

The Amazon Chime member ID (user ID or bot ID).

Destructuring the Response

data CreateRoomMembershipResponse

See: newCreateRoomMembershipResponse smart constructor.

Constructors

CreateRoomMembershipResponse'

Fields

roomMembership :: Maybe RoomMembership

The room membership details.

httpStatus :: Int

The response's http status code.

newCreateRoomMembershipResponse

Arguments

:: Int
$sel:httpStatus:CreateRoomMembershipResponse'

-> CreateRoomMembershipResponse

Create a value of CreateRoomMembershipResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:roomMembership:CreateRoomMembershipResponse', createRoomMembershipResponse_roomMembership - The room membership details.

$sel:httpStatus:CreateRoomMembershipResponse', createRoomMembershipResponse_httpStatus - The response's http status code.

Response Lenses

createRoomMembershipResponse_roomMembership :: Lens' CreateRoomMembershipResponse (Maybe RoomMembership)

The room membership details.

createRoomMembershipResponse_httpStatus :: Lens' CreateRoomMembershipResponse Int

The response's http status code.