Amazonka.Chime.GetVoiceConnectorLoggingConfiguration

Description

Retrieves the logging configuration details for the specified Amazon Chime Voice Connector. Shows whether SIP message logs are enabled for sending to Amazon CloudWatch Logs.

Creating a Request

data GetVoiceConnectorLoggingConfiguration

See: newGetVoiceConnectorLoggingConfiguration smart constructor.

Constructors

GetVoiceConnectorLoggingConfiguration'

Fields

voiceConnectorId :: Text

The Amazon Chime Voice Connector ID.

newGetVoiceConnectorLoggingConfiguration

Arguments

:: Text
$sel:voiceConnectorId:GetVoiceConnectorLoggingConfiguration'

-> GetVoiceConnectorLoggingConfiguration

Create a value of GetVoiceConnectorLoggingConfiguration with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:voiceConnectorId:GetVoiceConnectorLoggingConfiguration', getVoiceConnectorLoggingConfiguration_voiceConnectorId - The Amazon Chime Voice Connector ID.

Request Lenses

getVoiceConnectorLoggingConfiguration_voiceConnectorId :: Lens' GetVoiceConnectorLoggingConfiguration Text

The Amazon Chime Voice Connector ID.

Destructuring the Response

data GetVoiceConnectorLoggingConfigurationResponse

See: newGetVoiceConnectorLoggingConfigurationResponse smart constructor.

Constructors

GetVoiceConnectorLoggingConfigurationResponse'

Fields

loggingConfiguration :: Maybe LoggingConfiguration

The logging configuration details.

httpStatus :: Int

The response's http status code.

newGetVoiceConnectorLoggingConfigurationResponse

Arguments

:: Int
$sel:httpStatus:GetVoiceConnectorLoggingConfigurationResponse'

-> GetVoiceConnectorLoggingConfigurationResponse

Create a value of GetVoiceConnectorLoggingConfigurationResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:loggingConfiguration:GetVoiceConnectorLoggingConfigurationResponse', getVoiceConnectorLoggingConfigurationResponse_loggingConfiguration - The logging configuration details.

$sel:httpStatus:GetVoiceConnectorLoggingConfigurationResponse', getVoiceConnectorLoggingConfigurationResponse_httpStatus - The response's http status code.

Response Lenses

getVoiceConnectorLoggingConfigurationResponse_loggingConfiguration :: Lens' GetVoiceConnectorLoggingConfigurationResponse (Maybe LoggingConfiguration)

The logging configuration details.

getVoiceConnectorLoggingConfigurationResponse_httpStatus :: Lens' GetVoiceConnectorLoggingConfigurationResponse Int

The response's http status code.