Amazonka.Chime.InviteUsers

Description

Sends email to a maximum of 50 users, inviting them to the specified Amazon Chime Team account. Only Team account types are currently supported for this action.

Creating a Request

data InviteUsers

See: newInviteUsers smart constructor.

Constructors

InviteUsers'

Fields

userType :: Maybe UserType

The user type.

accountId :: Text

The Amazon Chime account ID.

userEmailList :: [Sensitive Text]

The user email addresses to which to send the email invitation.

newInviteUsers

Arguments

:: Text
$sel:accountId:InviteUsers'

-> InviteUsers

Create a value of InviteUsers with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:userType:InviteUsers', inviteUsers_userType - The user type.

$sel:accountId:InviteUsers', inviteUsers_accountId - The Amazon Chime account ID.

$sel:userEmailList:InviteUsers', inviteUsers_userEmailList - The user email addresses to which to send the email invitation.

Request Lenses

inviteUsers_userType :: Lens' InviteUsers (Maybe UserType)

The user type.

inviteUsers_accountId :: Lens' InviteUsers Text

The Amazon Chime account ID.

inviteUsers_userEmailList :: Lens' InviteUsers [Text]

The user email addresses to which to send the email invitation.

Destructuring the Response

data InviteUsersResponse

See: newInviteUsersResponse smart constructor.

Constructors

InviteUsersResponse'

Fields

invites :: Maybe [Invite]

The email invitation details.

httpStatus :: Int

The response's http status code.

newInviteUsersResponse

Arguments

:: Int
$sel:httpStatus:InviteUsersResponse'

-> InviteUsersResponse

Create a value of InviteUsersResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:invites:InviteUsersResponse', inviteUsersResponse_invites - The email invitation details.

$sel:httpStatus:InviteUsersResponse', inviteUsersResponse_httpStatus - The response's http status code.

Response Lenses

inviteUsersResponse_invites :: Lens' InviteUsersResponse (Maybe [Invite])

The email invitation details.

inviteUsersResponse_httpStatus :: Lens' InviteUsersResponse Int

The response's http status code.