Amazonka.Chime.ListAttendeeTags

Description

Lists the tags applied to an Amazon Chime SDK attendee resource.

Creating a Request

data ListAttendeeTags

See: newListAttendeeTags smart constructor.

Constructors

ListAttendeeTags'

Fields

meetingId :: Text

The Amazon Chime SDK meeting ID.

attendeeId :: Text

The Amazon Chime SDK attendee ID.

newListAttendeeTags

Arguments

:: Text
$sel:meetingId:ListAttendeeTags'

-> Text
$sel:attendeeId:ListAttendeeTags'

-> ListAttendeeTags

Create a value of ListAttendeeTags with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:meetingId:ListAttendeeTags', listAttendeeTags_meetingId - The Amazon Chime SDK meeting ID.

$sel:attendeeId:ListAttendeeTags', listAttendeeTags_attendeeId - The Amazon Chime SDK attendee ID.

Request Lenses

listAttendeeTags_meetingId :: Lens' ListAttendeeTags Text

The Amazon Chime SDK meeting ID.

listAttendeeTags_attendeeId :: Lens' ListAttendeeTags Text

The Amazon Chime SDK attendee ID.

Destructuring the Response

data ListAttendeeTagsResponse

See: newListAttendeeTagsResponse smart constructor.

Constructors

ListAttendeeTagsResponse'

Fields

tags :: Maybe (NonEmpty Tag)

A list of tag key-value pairs.

httpStatus :: Int

The response's http status code.

newListAttendeeTagsResponse

Arguments

:: Int
$sel:httpStatus:ListAttendeeTagsResponse'

-> ListAttendeeTagsResponse

Create a value of ListAttendeeTagsResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:tags:ListAttendeeTagsResponse', listAttendeeTagsResponse_tags - A list of tag key-value pairs.

$sel:httpStatus:ListAttendeeTagsResponse', listAttendeeTagsResponse_httpStatus - The response's http status code.

Response Lenses

listAttendeeTagsResponse_tags :: Lens' ListAttendeeTagsResponse (Maybe (NonEmpty Tag))

A list of tag key-value pairs.

listAttendeeTagsResponse_httpStatus :: Lens' ListAttendeeTagsResponse Int

The response's http status code.