Amazonka.Chime.ListProxySessions

Description

Lists the proxy sessions for the specified Amazon Chime Voice Connector.

Creating a Request

data ListProxySessions

See: newListProxySessions smart constructor.

Constructors

ListProxySessions'

Fields

status :: Maybe ProxySessionStatus

The proxy session status.

nextToken :: Maybe Text

The token to use to retrieve the next page of results.

maxResults :: Maybe Natural

The maximum number of results to return in a single call.

voiceConnectorId :: Text

The Amazon Chime voice connector ID.

newListProxySessions

Arguments

:: Text
$sel:voiceConnectorId:ListProxySessions'

-> ListProxySessions

Create a value of ListProxySessions with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:status:ListProxySessions', listProxySessions_status - The proxy session status.

$sel:nextToken:ListProxySessions', listProxySessions_nextToken - The token to use to retrieve the next page of results.

$sel:maxResults:ListProxySessions', listProxySessions_maxResults - The maximum number of results to return in a single call.

$sel:voiceConnectorId:ListProxySessions', listProxySessions_voiceConnectorId - The Amazon Chime voice connector ID.

Request Lenses

listProxySessions_status :: Lens' ListProxySessions (Maybe ProxySessionStatus)

The proxy session status.

listProxySessions_nextToken :: Lens' ListProxySessions (Maybe Text)

The token to use to retrieve the next page of results.

listProxySessions_maxResults :: Lens' ListProxySessions (Maybe Natural)

The maximum number of results to return in a single call.

listProxySessions_voiceConnectorId :: Lens' ListProxySessions Text

The Amazon Chime voice connector ID.

Destructuring the Response

data ListProxySessionsResponse

See: newListProxySessionsResponse smart constructor.

Constructors

ListProxySessionsResponse'

Fields

nextToken :: Maybe Text

The token to use to retrieve the next page of results.

proxySessions :: Maybe [ProxySession]

The proxy session details.

httpStatus :: Int

The response's http status code.

newListProxySessionsResponse

Arguments

:: Int
$sel:httpStatus:ListProxySessionsResponse'

-> ListProxySessionsResponse

Create a value of ListProxySessionsResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:nextToken:ListProxySessions', listProxySessionsResponse_nextToken - The token to use to retrieve the next page of results.

$sel:proxySessions:ListProxySessionsResponse', listProxySessionsResponse_proxySessions - The proxy session details.

$sel:httpStatus:ListProxySessionsResponse', listProxySessionsResponse_httpStatus - The response's http status code.

Response Lenses

listProxySessionsResponse_nextToken :: Lens' ListProxySessionsResponse (Maybe Text)

The token to use to retrieve the next page of results.

listProxySessionsResponse_proxySessions :: Lens' ListProxySessionsResponse (Maybe [ProxySession])

The proxy session details.

listProxySessionsResponse_httpStatus :: Lens' ListProxySessionsResponse Int

The response's http status code.