Amazonka.Chime.ListRooms

Description

Lists the room details for the specified Amazon Chime Enterprise account. Optionally, filter the results by a member ID (user ID or bot ID) to see a list of rooms that the member belongs to.

Creating a Request

data ListRooms

See: newListRooms smart constructor.

Constructors

ListRooms'

Fields

memberId :: Maybe Text

The member ID (user ID or bot ID).

nextToken :: Maybe Text

The token to use to retrieve the next page of results.

maxResults :: Maybe Natural

The maximum number of results to return in a single call.

accountId :: Text

The Amazon Chime account ID.

newListRooms

Arguments

:: Text
$sel:accountId:ListRooms'

-> ListRooms

Create a value of ListRooms with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:memberId:ListRooms', listRooms_memberId - The member ID (user ID or bot ID).

$sel:nextToken:ListRooms', listRooms_nextToken - The token to use to retrieve the next page of results.

$sel:maxResults:ListRooms', listRooms_maxResults - The maximum number of results to return in a single call.

$sel:accountId:ListRooms', listRooms_accountId - The Amazon Chime account ID.

Request Lenses

listRooms_memberId :: Lens' ListRooms (Maybe Text)

The member ID (user ID or bot ID).

listRooms_nextToken :: Lens' ListRooms (Maybe Text)

The token to use to retrieve the next page of results.

listRooms_maxResults :: Lens' ListRooms (Maybe Natural)

The maximum number of results to return in a single call.

listRooms_accountId :: Lens' ListRooms Text

The Amazon Chime account ID.

Destructuring the Response

data ListRoomsResponse

See: newListRoomsResponse smart constructor.

Constructors

ListRoomsResponse'

Fields

rooms :: Maybe [Room]

The room details.

nextToken :: Maybe Text

The token to use to retrieve the next page of results.

httpStatus :: Int

The response's http status code.

newListRoomsResponse

Arguments

:: Int
$sel:httpStatus:ListRoomsResponse'

-> ListRoomsResponse

Create a value of ListRoomsResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:rooms:ListRoomsResponse', listRoomsResponse_rooms - The room details.

$sel:nextToken:ListRooms', listRoomsResponse_nextToken - The token to use to retrieve the next page of results.

$sel:httpStatus:ListRoomsResponse', listRoomsResponse_httpStatus - The response's http status code.

Response Lenses

listRoomsResponse_rooms :: Lens' ListRoomsResponse (Maybe [Room])

The room details.

listRoomsResponse_nextToken :: Lens' ListRoomsResponse (Maybe Text)

The token to use to retrieve the next page of results.

listRoomsResponse_httpStatus :: Lens' ListRoomsResponse Int

The response's http status code.