Amazonka.Chime.PutVoiceConnectorEmergencyCallingConfiguration

Description

Puts emergency calling configuration details to the specified Amazon Chime Voice Connector, such as emergency phone numbers and calling countries. Origination and termination settings must be enabled for the Amazon Chime Voice Connector before emergency calling can be configured.

Creating a Request

data PutVoiceConnectorEmergencyCallingConfiguration

See: newPutVoiceConnectorEmergencyCallingConfiguration smart constructor.

Constructors

PutVoiceConnectorEmergencyCallingConfiguration'

Fields

voiceConnectorId :: Text

The Amazon Chime Voice Connector ID.

emergencyCallingConfiguration :: EmergencyCallingConfiguration

The emergency calling configuration details.

newPutVoiceConnectorEmergencyCallingConfiguration

Arguments

:: Text
$sel:voiceConnectorId:PutVoiceConnectorEmergencyCallingConfiguration'

-> EmergencyCallingConfiguration
$sel:emergencyCallingConfiguration:PutVoiceConnectorEmergencyCallingConfiguration'

-> PutVoiceConnectorEmergencyCallingConfiguration

Create a value of PutVoiceConnectorEmergencyCallingConfiguration with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:voiceConnectorId:PutVoiceConnectorEmergencyCallingConfiguration', putVoiceConnectorEmergencyCallingConfiguration_voiceConnectorId - The Amazon Chime Voice Connector ID.

$sel:emergencyCallingConfiguration:PutVoiceConnectorEmergencyCallingConfiguration', putVoiceConnectorEmergencyCallingConfiguration_emergencyCallingConfiguration - The emergency calling configuration details.

Request Lenses

putVoiceConnectorEmergencyCallingConfiguration_voiceConnectorId :: Lens' PutVoiceConnectorEmergencyCallingConfiguration Text

The Amazon Chime Voice Connector ID.

putVoiceConnectorEmergencyCallingConfiguration_emergencyCallingConfiguration :: Lens' PutVoiceConnectorEmergencyCallingConfiguration EmergencyCallingConfiguration

The emergency calling configuration details.

Destructuring the Response

data PutVoiceConnectorEmergencyCallingConfigurationResponse

See: newPutVoiceConnectorEmergencyCallingConfigurationResponse smart constructor.

Constructors

PutVoiceConnectorEmergencyCallingConfigurationResponse'

Fields

emergencyCallingConfiguration :: Maybe EmergencyCallingConfiguration

The emergency calling configuration details.

httpStatus :: Int

The response's http status code.

newPutVoiceConnectorEmergencyCallingConfigurationResponse

Arguments

:: Int
$sel:httpStatus:PutVoiceConnectorEmergencyCallingConfigurationResponse'

-> PutVoiceConnectorEmergencyCallingConfigurationResponse

Create a value of PutVoiceConnectorEmergencyCallingConfigurationResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:emergencyCallingConfiguration:PutVoiceConnectorEmergencyCallingConfiguration', putVoiceConnectorEmergencyCallingConfigurationResponse_emergencyCallingConfiguration - The emergency calling configuration details.

$sel:httpStatus:PutVoiceConnectorEmergencyCallingConfigurationResponse', putVoiceConnectorEmergencyCallingConfigurationResponse_httpStatus - The response's http status code.

Response Lenses

putVoiceConnectorEmergencyCallingConfigurationResponse_emergencyCallingConfiguration :: Lens' PutVoiceConnectorEmergencyCallingConfigurationResponse (Maybe EmergencyCallingConfiguration)

The emergency calling configuration details.

putVoiceConnectorEmergencyCallingConfigurationResponse_httpStatus :: Lens' PutVoiceConnectorEmergencyCallingConfigurationResponse Int

The response's http status code.