Amazonka.Chime.PutVoiceConnectorProxy

Description

Puts the specified proxy configuration to the specified Amazon Chime Voice Connector.

Creating a Request

data PutVoiceConnectorProxy

See: newPutVoiceConnectorProxy smart constructor.

Constructors

PutVoiceConnectorProxy'

Fields

disabled :: Maybe Bool

When true, stops proxy sessions from being created on the specified Amazon Chime Voice Connector.

fallBackPhoneNumber :: Maybe (Sensitive Text)

The phone number to route calls to after a proxy session expires.

defaultSessionExpiryMinutes :: Int

The default number of minutes allowed for proxy sessions.

phoneNumberPoolCountries :: NonEmpty Text

The countries for proxy phone numbers to be selected from.

voiceConnectorId :: Text

The Amazon Chime voice connector ID.

newPutVoiceConnectorProxy

Arguments

:: Int
$sel:defaultSessionExpiryMinutes:PutVoiceConnectorProxy'

-> NonEmpty Text
$sel:phoneNumberPoolCountries:PutVoiceConnectorProxy'

-> Text
$sel:voiceConnectorId:PutVoiceConnectorProxy'

-> PutVoiceConnectorProxy

Create a value of PutVoiceConnectorProxy with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:disabled:PutVoiceConnectorProxy', putVoiceConnectorProxy_disabled - When true, stops proxy sessions from being created on the specified Amazon Chime Voice Connector.

$sel:fallBackPhoneNumber:PutVoiceConnectorProxy', putVoiceConnectorProxy_fallBackPhoneNumber - The phone number to route calls to after a proxy session expires.

$sel:defaultSessionExpiryMinutes:PutVoiceConnectorProxy', putVoiceConnectorProxy_defaultSessionExpiryMinutes - The default number of minutes allowed for proxy sessions.

$sel:phoneNumberPoolCountries:PutVoiceConnectorProxy', putVoiceConnectorProxy_phoneNumberPoolCountries - The countries for proxy phone numbers to be selected from.

$sel:voiceConnectorId:PutVoiceConnectorProxy', putVoiceConnectorProxy_voiceConnectorId - The Amazon Chime voice connector ID.

Request Lenses

putVoiceConnectorProxy_disabled :: Lens' PutVoiceConnectorProxy (Maybe Bool)

When true, stops proxy sessions from being created on the specified Amazon Chime Voice Connector.

putVoiceConnectorProxy_fallBackPhoneNumber :: Lens' PutVoiceConnectorProxy (Maybe Text)

The phone number to route calls to after a proxy session expires.

putVoiceConnectorProxy_defaultSessionExpiryMinutes :: Lens' PutVoiceConnectorProxy Int

The default number of minutes allowed for proxy sessions.

putVoiceConnectorProxy_phoneNumberPoolCountries :: Lens' PutVoiceConnectorProxy (NonEmpty Text)

The countries for proxy phone numbers to be selected from.

putVoiceConnectorProxy_voiceConnectorId :: Lens' PutVoiceConnectorProxy Text

The Amazon Chime voice connector ID.

Destructuring the Response

data PutVoiceConnectorProxyResponse

See: newPutVoiceConnectorProxyResponse smart constructor.

Constructors

PutVoiceConnectorProxyResponse'

Fields

proxy :: Maybe Proxy

The proxy configuration details.

httpStatus :: Int

The response's http status code.

newPutVoiceConnectorProxyResponse

Arguments

:: Int
$sel:httpStatus:PutVoiceConnectorProxyResponse'

-> PutVoiceConnectorProxyResponse

Create a value of PutVoiceConnectorProxyResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:proxy:PutVoiceConnectorProxyResponse', putVoiceConnectorProxyResponse_proxy - The proxy configuration details.

$sel:httpStatus:PutVoiceConnectorProxyResponse', putVoiceConnectorProxyResponse_httpStatus - The response's http status code.

Response Lenses

putVoiceConnectorProxyResponse_proxy :: Lens' PutVoiceConnectorProxyResponse (Maybe Proxy)

The proxy configuration details.

putVoiceConnectorProxyResponse_httpStatus :: Lens' PutVoiceConnectorProxyResponse Int

The response's http status code.