Amazonka.Chime.Types.Account

Description

Documentation

data Account

The Amazon Chime account details. An AWS account can have multiple Amazon Chime accounts.

See: newAccount smart constructor.

Constructors

Account'

Fields

signinDelegateGroups :: Maybe [SigninDelegateGroup]

The sign-in delegate groups associated with the account.

accountStatus :: Maybe AccountStatus

The status of the account.

defaultLicense :: Maybe License

The default license for the Amazon Chime account.

supportedLicenses :: Maybe [License]

Supported licenses for the Amazon Chime account.

createdTimestamp :: Maybe POSIX

The Amazon Chime account creation timestamp, in ISO 8601 format.

accountType :: Maybe AccountType

The Amazon Chime account type. For more information about different account types, see Managing Your Amazon Chime Accounts in the Amazon Chime Administration Guide.

awsAccountId :: Text

The AWS account ID.

accountId :: Text

The Amazon Chime account ID.

name :: Text

The Amazon Chime account name.

newAccount

Arguments

:: Text
$sel:awsAccountId:Account'

-> Text
$sel:accountId:Account'

-> Text
$sel:name:Account'

-> Account

Create a value of Account with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:signinDelegateGroups:Account', account_signinDelegateGroups - The sign-in delegate groups associated with the account.

$sel:accountStatus:Account', account_accountStatus - The status of the account.

$sel:defaultLicense:Account', account_defaultLicense - The default license for the Amazon Chime account.

$sel:supportedLicenses:Account', account_supportedLicenses - Supported licenses for the Amazon Chime account.

$sel:createdTimestamp:Account', account_createdTimestamp - The Amazon Chime account creation timestamp, in ISO 8601 format.

$sel:accountType:Account', account_accountType - The Amazon Chime account type. For more information about different account types, see Managing Your Amazon Chime Accounts in the Amazon Chime Administration Guide.

$sel:awsAccountId:Account', account_awsAccountId - The AWS account ID.

$sel:accountId:Account', account_accountId - The Amazon Chime account ID.

$sel:name:Account', account_name - The Amazon Chime account name.

account_signinDelegateGroups :: Lens' Account (Maybe [SigninDelegateGroup])

The sign-in delegate groups associated with the account.

account_accountStatus :: Lens' Account (Maybe AccountStatus)

The status of the account.

account_defaultLicense :: Lens' Account (Maybe License)

The default license for the Amazon Chime account.

account_supportedLicenses :: Lens' Account (Maybe [License])

Supported licenses for the Amazon Chime account.

account_createdTimestamp :: Lens' Account (Maybe UTCTime)

The Amazon Chime account creation timestamp, in ISO 8601 format.

account_accountType :: Lens' Account (Maybe AccountType)

The Amazon Chime account type. For more information about different account types, see Managing Your Amazon Chime Accounts in the Amazon Chime Administration Guide.

account_awsAccountId :: Lens' Account Text

The AWS account ID.

account_accountId :: Lens' Account Text

The Amazon Chime account ID.

account_name :: Lens' Account Text

The Amazon Chime account name.