Amazonka.Chime.Types.Invite

Description

Documentation

data Invite

Invitation object returned after emailing users to invite them to join the Amazon Chime Team account.

See: newInvite smart constructor.

Constructors

Invite'

Fields

status :: Maybe InviteStatus

The status of the invite.

emailStatus :: Maybe EmailStatus

The status of the invite email.

inviteId :: Maybe Text

The invite ID.

emailAddress :: Maybe (Sensitive Text)

The email address to which the invite is sent.

newInvite :: Invite

Create a value of Invite with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:status:Invite', invite_status - The status of the invite.

$sel:emailStatus:Invite', invite_emailStatus - The status of the invite email.

$sel:inviteId:Invite', invite_inviteId - The invite ID.

$sel:emailAddress:Invite', invite_emailAddress - The email address to which the invite is sent.

invite_status :: Lens' Invite (Maybe InviteStatus)

The status of the invite.

invite_emailStatus :: Lens' Invite (Maybe EmailStatus)

The status of the invite email.

invite_inviteId :: Lens' Invite (Maybe Text)

The invite ID.

invite_emailAddress :: Lens' Invite (Maybe Text)

The email address to which the invite is sent.