Amazonka.Chime.Types.Meeting

Description

Documentation

data Meeting

A meeting created using the Amazon Chime SDK.

See: newMeeting smart constructor.

Constructors

Meeting'

Fields

mediaRegion :: Maybe Text

The Region in which you create the meeting. Available values: af-south-1, ap-northeast-1, ap-northeast-2, ap-south-1, ap-southeast-1, ap-southeast-2, ca-central-1, eu-central-1, eu-north-1, eu-south-1, eu-west-1, eu-west-2, eu-west-3, sa-east-1, us-east-1, us-east-2, us-west-1, us-west-2.

mediaPlacement :: Maybe MediaPlacement

The media placement for the meeting.

externalMeetingId :: Maybe (Sensitive Text)

The external meeting ID.

meetingId :: Maybe Text

The Amazon Chime SDK meeting ID.

newMeeting :: Meeting

Create a value of Meeting with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:mediaRegion:Meeting', meeting_mediaRegion - The Region in which you create the meeting. Available values: af-south-1, ap-northeast-1, ap-northeast-2, ap-south-1, ap-southeast-1, ap-southeast-2, ca-central-1, eu-central-1, eu-north-1, eu-south-1, eu-west-1, eu-west-2, eu-west-3, sa-east-1, us-east-1, us-east-2, us-west-1, us-west-2.

$sel:mediaPlacement:Meeting', meeting_mediaPlacement - The media placement for the meeting.

$sel:externalMeetingId:Meeting', meeting_externalMeetingId - The external meeting ID.

$sel:meetingId:Meeting', meeting_meetingId - The Amazon Chime SDK meeting ID.

meeting_mediaRegion :: Lens' Meeting (Maybe Text)

The Region in which you create the meeting. Available values: af-south-1, ap-northeast-1, ap-northeast-2, ap-south-1, ap-southeast-1, ap-southeast-2, ca-central-1, eu-central-1, eu-north-1, eu-south-1, eu-west-1, eu-west-2, eu-west-3, sa-east-1, us-east-1, us-east-2, us-west-1, us-west-2.

meeting_mediaPlacement :: Lens' Meeting (Maybe MediaPlacement)

The media placement for the meeting.

meeting_externalMeetingId :: Lens' Meeting (Maybe Text)

The external meeting ID.

meeting_meetingId :: Lens' Meeting (Maybe Text)

The Amazon Chime SDK meeting ID.