Amazonka.Chime.Types.Member

Description

Documentation

data Member

The member details, such as email address, name, member ID, and member type.

See: newMember smart constructor.

Constructors

Member'

Fields

fullName :: Maybe (Sensitive Text)

The member name.

email :: Maybe (Sensitive Text)

The member email address.

memberId :: Maybe Text

The member ID (user ID or bot ID).

memberType :: Maybe MemberType

The member type.

accountId :: Maybe Text

The Amazon Chime account ID.

newMember :: Member

Create a value of Member with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:fullName:Member', member_fullName - The member name.

$sel:email:Member', member_email - The member email address.

$sel:memberId:Member', member_memberId - The member ID (user ID or bot ID).

$sel:memberType:Member', member_memberType - The member type.

$sel:accountId:Member', member_accountId - The Amazon Chime account ID.

member_fullName :: Lens' Member (Maybe Text)

The member name.

member_email :: Lens' Member (Maybe Text)

The member email address.

member_memberId :: Lens' Member (Maybe Text)

The member ID (user ID or bot ID).

member_memberType :: Lens' Member (Maybe MemberType)

The member type.

member_accountId :: Lens' Member (Maybe Text)

The Amazon Chime account ID.