Amazonka.Chime.Types.PhoneNumberError

Description

Documentation

data PhoneNumberError

If the phone number action fails for one or more of the phone numbers in the request, a list of the phone numbers is returned, along with error codes and error messages.

See: newPhoneNumberError smart constructor.

Constructors

PhoneNumberError'

Fields

phoneNumberId :: Maybe Text

The phone number ID for which the action failed.

errorCode :: Maybe ErrorCode

The error code.

errorMessage :: Maybe Text

The error message.

newPhoneNumberError :: PhoneNumberError

Create a value of PhoneNumberError with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:phoneNumberId:PhoneNumberError', phoneNumberError_phoneNumberId - The phone number ID for which the action failed.

$sel:errorCode:PhoneNumberError', phoneNumberError_errorCode - The error code.

$sel:errorMessage:PhoneNumberError', phoneNumberError_errorMessage - The error message.

phoneNumberError_phoneNumberId :: Lens' PhoneNumberError (Maybe Text)

The phone number ID for which the action failed.

phoneNumberError_errorCode :: Lens' PhoneNumberError (Maybe ErrorCode)

The error code.

phoneNumberError_errorMessage :: Lens' PhoneNumberError (Maybe Text)

The error message.