Amazonka.Chime.Types.PhoneNumberOrder

Description

Documentation

data PhoneNumberOrder

The details of a phone number order created for Amazon Chime.

See: newPhoneNumberOrder smart constructor.

Constructors

PhoneNumberOrder'

Fields

status :: Maybe PhoneNumberOrderStatus

The status of the phone number order.

orderedPhoneNumbers :: Maybe [OrderedPhoneNumber]

The ordered phone number details, such as the phone number in E.164 format and the phone number status.

updatedTimestamp :: Maybe POSIX

The updated phone number order time stamp, in ISO 8601 format.

productType :: Maybe PhoneNumberProductType

The phone number order product type.

phoneNumberOrderId :: Maybe Text

The phone number order ID.

createdTimestamp :: Maybe POSIX

The phone number order creation time stamp, in ISO 8601 format.

newPhoneNumberOrder :: PhoneNumberOrder

Create a value of PhoneNumberOrder with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:status:PhoneNumberOrder', phoneNumberOrder_status - The status of the phone number order.

$sel:orderedPhoneNumbers:PhoneNumberOrder', phoneNumberOrder_orderedPhoneNumbers - The ordered phone number details, such as the phone number in E.164 format and the phone number status.

$sel:updatedTimestamp:PhoneNumberOrder', phoneNumberOrder_updatedTimestamp - The updated phone number order time stamp, in ISO 8601 format.

$sel:productType:PhoneNumberOrder', phoneNumberOrder_productType - The phone number order product type.

$sel:phoneNumberOrderId:PhoneNumberOrder', phoneNumberOrder_phoneNumberOrderId - The phone number order ID.

$sel:createdTimestamp:PhoneNumberOrder', phoneNumberOrder_createdTimestamp - The phone number order creation time stamp, in ISO 8601 format.

phoneNumberOrder_status :: Lens' PhoneNumberOrder (Maybe PhoneNumberOrderStatus)

The status of the phone number order.

phoneNumberOrder_orderedPhoneNumbers :: Lens' PhoneNumberOrder (Maybe [OrderedPhoneNumber])

The ordered phone number details, such as the phone number in E.164 format and the phone number status.

phoneNumberOrder_updatedTimestamp :: Lens' PhoneNumberOrder (Maybe UTCTime)

The updated phone number order time stamp, in ISO 8601 format.

phoneNumberOrder_productType :: Lens' PhoneNumberOrder (Maybe PhoneNumberProductType)

The phone number order product type.

phoneNumberOrder_phoneNumberOrderId :: Lens' PhoneNumberOrder (Maybe Text)

The phone number order ID.

phoneNumberOrder_createdTimestamp :: Lens' PhoneNumberOrder (Maybe UTCTime)

The phone number order creation time stamp, in ISO 8601 format.