Amazonka.Chime.Types.Room

Description

Documentation

data Room

The Amazon Chime chat room details.

See: newRoom smart constructor.

Constructors

Room'

Fields

updatedTimestamp :: Maybe POSIX

The room update timestamp, in ISO 8601 format.

createdBy :: Maybe Text

The identifier of the room creator.

accountId :: Maybe Text

The Amazon Chime account ID.

name :: Maybe (Sensitive Text)

The room name.

roomId :: Maybe Text

The room ID.

createdTimestamp :: Maybe POSIX

The room creation timestamp, in ISO 8601 format.

newRoom :: Room

Create a value of Room with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:updatedTimestamp:Room', room_updatedTimestamp - The room update timestamp, in ISO 8601 format.

$sel:createdBy:Room', room_createdBy - The identifier of the room creator.

$sel:accountId:Room', room_accountId - The Amazon Chime account ID.

$sel:name:Room', room_name - The room name.

$sel:roomId:Room', room_roomId - The room ID.

$sel:createdTimestamp:Room', room_createdTimestamp - The room creation timestamp, in ISO 8601 format.

room_updatedTimestamp :: Lens' Room (Maybe UTCTime)

The room update timestamp, in ISO 8601 format.

room_createdBy :: Lens' Room (Maybe Text)

The identifier of the room creator.

room_accountId :: Lens' Room (Maybe Text)

The Amazon Chime account ID.

room_name :: Lens' Room (Maybe Text)

The room name.

room_roomId :: Lens' Room (Maybe Text)

The room ID.

room_createdTimestamp :: Lens' Room (Maybe UTCTime)

The room creation timestamp, in ISO 8601 format.