Amazonka.Chime.Types.UpdatePhoneNumberRequestItem

Description

Documentation

data UpdatePhoneNumberRequestItem

The phone number ID, product type, or calling name fields to update, used with the BatchUpdatePhoneNumber and UpdatePhoneNumber actions.

See: newUpdatePhoneNumberRequestItem smart constructor.

Constructors

UpdatePhoneNumberRequestItem'

Fields

productType :: Maybe PhoneNumberProductType

The product type to update.

callingName :: Maybe (Sensitive Text)

The outbound calling name to update.

phoneNumberId :: Text

The phone number ID to update.

newUpdatePhoneNumberRequestItem

Arguments

:: Text
$sel:phoneNumberId:UpdatePhoneNumberRequestItem'

-> UpdatePhoneNumberRequestItem

Create a value of UpdatePhoneNumberRequestItem with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:productType:UpdatePhoneNumberRequestItem', updatePhoneNumberRequestItem_productType - The product type to update.

$sel:callingName:UpdatePhoneNumberRequestItem', updatePhoneNumberRequestItem_callingName - The outbound calling name to update.

$sel:phoneNumberId:UpdatePhoneNumberRequestItem', updatePhoneNumberRequestItem_phoneNumberId - The phone number ID to update.

updatePhoneNumberRequestItem_productType :: Lens' UpdatePhoneNumberRequestItem (Maybe PhoneNumberProductType)

The product type to update.

updatePhoneNumberRequestItem_callingName :: Lens' UpdatePhoneNumberRequestItem (Maybe Text)

The outbound calling name to update.

updatePhoneNumberRequestItem_phoneNumberId :: Lens' UpdatePhoneNumberRequestItem Text

The phone number ID to update.