Amazonka.Chime.UpdateRoomMembership

Description

Updates room membership details, such as the member role, for a room in an Amazon Chime Enterprise account. The member role designates whether the member is a chat room administrator or a general chat room member. The member role can be updated only for user IDs.

Creating a Request

data UpdateRoomMembership

See: newUpdateRoomMembership smart constructor.

Constructors

UpdateRoomMembership'

Fields

role' :: Maybe RoomMembershipRole

The role of the member.

accountId :: Text

The Amazon Chime account ID.

roomId :: Text

The room ID.

memberId :: Text

The member ID.

newUpdateRoomMembership

Arguments

:: Text
$sel:accountId:UpdateRoomMembership'

-> Text
$sel:roomId:UpdateRoomMembership'

-> Text
$sel:memberId:UpdateRoomMembership'

-> UpdateRoomMembership

Create a value of UpdateRoomMembership with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:role':UpdateRoomMembership', updateRoomMembership_role - The role of the member.

$sel:accountId:UpdateRoomMembership', updateRoomMembership_accountId - The Amazon Chime account ID.

$sel:roomId:UpdateRoomMembership', updateRoomMembership_roomId - The room ID.

$sel:memberId:UpdateRoomMembership', updateRoomMembership_memberId - The member ID.

Request Lenses

updateRoomMembership_role :: Lens' UpdateRoomMembership (Maybe RoomMembershipRole)

The role of the member.

updateRoomMembership_accountId :: Lens' UpdateRoomMembership Text

The Amazon Chime account ID.

updateRoomMembership_roomId :: Lens' UpdateRoomMembership Text

The room ID.

updateRoomMembership_memberId :: Lens' UpdateRoomMembership Text

The member ID.

Destructuring the Response

data UpdateRoomMembershipResponse

See: newUpdateRoomMembershipResponse smart constructor.

Constructors

UpdateRoomMembershipResponse'

Fields

roomMembership :: Maybe RoomMembership

The room membership details.

httpStatus :: Int

The response's http status code.

newUpdateRoomMembershipResponse

Arguments

:: Int
$sel:httpStatus:UpdateRoomMembershipResponse'

-> UpdateRoomMembershipResponse

Create a value of UpdateRoomMembershipResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:roomMembership:UpdateRoomMembershipResponse', updateRoomMembershipResponse_roomMembership - The room membership details.

$sel:httpStatus:UpdateRoomMembershipResponse', updateRoomMembershipResponse_httpStatus - The response's http status code.

Response Lenses

updateRoomMembershipResponse_roomMembership :: Lens' UpdateRoomMembershipResponse (Maybe RoomMembership)

The room membership details.

updateRoomMembershipResponse_httpStatus :: Lens' UpdateRoomMembershipResponse Int

The response's http status code.