Amazonka.ChimeSDKIdentity.Lens

Description

Operations

CreateAppInstance

createAppInstance_metadata :: Lens' CreateAppInstance (Maybe Text)

The metadata of the AppInstance. Limited to a 1KB string in UTF-8.

createAppInstance_tags :: Lens' CreateAppInstance (Maybe (NonEmpty Tag))

Tags assigned to the AppInstanceUser.

createAppInstance_name :: Lens' CreateAppInstance Text

The name of the AppInstance.

createAppInstance_clientRequestToken :: Lens' CreateAppInstance Text

The ClientRequestToken of the AppInstance.

createAppInstanceResponse_appInstanceArn :: Lens' CreateAppInstanceResponse (Maybe Text)

The Amazon Resource Number (ARN) of the AppInstance.

createAppInstanceResponse_httpStatus :: Lens' CreateAppInstanceResponse Int

The response's http status code.

GetAppInstanceRetentionSettings

getAppInstanceRetentionSettings_appInstanceArn :: Lens' GetAppInstanceRetentionSettings Text

The ARN of the AppInstance.

getAppInstanceRetentionSettingsResponse_appInstanceRetentionSettings :: Lens' GetAppInstanceRetentionSettingsResponse (Maybe AppInstanceRetentionSettings)

The retention settings for the AppInstance.

getAppInstanceRetentionSettingsResponse_initiateDeletionTimestamp :: Lens' GetAppInstanceRetentionSettingsResponse (Maybe UTCTime)

The timestamp representing the time at which the specified items are retained, in Epoch Seconds.

getAppInstanceRetentionSettingsResponse_httpStatus :: Lens' GetAppInstanceRetentionSettingsResponse Int

The response's http status code.

DescribeAppInstanceAdmin

describeAppInstanceAdmin_appInstanceAdminArn :: Lens' DescribeAppInstanceAdmin Text

The ARN of the AppInstanceAdmin.

describeAppInstanceAdmin_appInstanceArn :: Lens' DescribeAppInstanceAdmin Text

The ARN of the AppInstance.

describeAppInstanceAdminResponse_appInstanceAdmin :: Lens' DescribeAppInstanceAdminResponse (Maybe AppInstanceAdmin)

The ARN and name of the AppInstanceUser, the ARN of the AppInstance, and the created and last-updated timestamps. All timestamps use epoch milliseconds.

describeAppInstanceAdminResponse_httpStatus :: Lens' DescribeAppInstanceAdminResponse Int

The response's http status code.

CreateAppInstanceUser

createAppInstanceUser_metadata :: Lens' CreateAppInstanceUser (Maybe Text)

The request's metadata. Limited to a 1KB string in UTF-8.

createAppInstanceUser_tags :: Lens' CreateAppInstanceUser (Maybe (NonEmpty Tag))

Tags assigned to the AppInstanceUser.

createAppInstanceUser_appInstanceArn :: Lens' CreateAppInstanceUser Text

The ARN of the AppInstance request.

createAppInstanceUser_appInstanceUserId :: Lens' CreateAppInstanceUser Text

The user ID of the AppInstance.

createAppInstanceUser_name :: Lens' CreateAppInstanceUser Text

The user's name.

createAppInstanceUser_clientRequestToken :: Lens' CreateAppInstanceUser Text

The token assigned to the user requesting an AppInstance.

createAppInstanceUserResponse_appInstanceUserArn :: Lens' CreateAppInstanceUserResponse (Maybe Text)

The user's ARN.

createAppInstanceUserResponse_httpStatus :: Lens' CreateAppInstanceUserResponse Int

The response's http status code.

PutAppInstanceRetentionSettings

putAppInstanceRetentionSettings_appInstanceArn :: Lens' PutAppInstanceRetentionSettings Text

The ARN of the AppInstance.

putAppInstanceRetentionSettings_appInstanceRetentionSettings :: Lens' PutAppInstanceRetentionSettings AppInstanceRetentionSettings

The time in days to retain data. Data type: number.

putAppInstanceRetentionSettingsResponse_appInstanceRetentionSettings :: Lens' PutAppInstanceRetentionSettingsResponse (Maybe AppInstanceRetentionSettings)

The time in days to retain data. Data type: number.

putAppInstanceRetentionSettingsResponse_initiateDeletionTimestamp :: Lens' PutAppInstanceRetentionSettingsResponse (Maybe UTCTime)

The time at which the API deletes data.

putAppInstanceRetentionSettingsResponse_httpStatus :: Lens' PutAppInstanceRetentionSettingsResponse Int

The response's http status code.

CreateAppInstanceAdmin

createAppInstanceAdmin_appInstanceAdminArn :: Lens' CreateAppInstanceAdmin Text

The ARN of the administrator of the current AppInstance.

createAppInstanceAdmin_appInstanceArn :: Lens' CreateAppInstanceAdmin Text

The ARN of the AppInstance.

createAppInstanceAdminResponse_appInstanceAdmin :: Lens' CreateAppInstanceAdminResponse (Maybe Identity)

The name and ARN of the admin for the AppInstance.

createAppInstanceAdminResponse_appInstanceArn :: Lens' CreateAppInstanceAdminResponse (Maybe Text)

The ARN of the of the admin for the AppInstance.

createAppInstanceAdminResponse_httpStatus :: Lens' CreateAppInstanceAdminResponse Int

The response's http status code.

ListAppInstanceAdmins

listAppInstanceAdmins_nextToken :: Lens' ListAppInstanceAdmins (Maybe Text)

The token returned from previous API requests until the number of administrators is reached.

listAppInstanceAdmins_maxResults :: Lens' ListAppInstanceAdmins (Maybe Natural)

The maximum number of administrators that you want to return.

listAppInstanceAdmins_appInstanceArn :: Lens' ListAppInstanceAdmins Text

The ARN of the AppInstance.

listAppInstanceAdminsResponse_nextToken :: Lens' ListAppInstanceAdminsResponse (Maybe Text)

The token returned from previous API requests until the number of administrators is reached.

listAppInstanceAdminsResponse_appInstanceAdmins :: Lens' ListAppInstanceAdminsResponse (Maybe [AppInstanceAdminSummary])

The information for each administrator.

listAppInstanceAdminsResponse_appInstanceArn :: Lens' ListAppInstanceAdminsResponse (Maybe Text)

The ARN of the AppInstance.

listAppInstanceAdminsResponse_httpStatus :: Lens' ListAppInstanceAdminsResponse Int

The response's http status code.

ListAppInstances

listAppInstances_nextToken :: Lens' ListAppInstances (Maybe Text)

The token passed by previous API requests until you reach the maximum number of AppInstances.

listAppInstances_maxResults :: Lens' ListAppInstances (Maybe Natural)

The maximum number of AppInstances that you want to return.

listAppInstancesResponse_appInstances :: Lens' ListAppInstancesResponse (Maybe [AppInstanceSummary])

The information for each AppInstance.

listAppInstancesResponse_nextToken :: Lens' ListAppInstancesResponse (Maybe Text)

The token passed by previous API requests until the maximum number of AppInstances is reached.

listAppInstancesResponse_httpStatus :: Lens' ListAppInstancesResponse Int

The response's http status code.

DescribeAppInstanceUser

describeAppInstanceUser_appInstanceUserArn :: Lens' DescribeAppInstanceUser Text

The ARN of the AppInstanceUser.

describeAppInstanceUserResponse_appInstanceUser :: Lens' DescribeAppInstanceUserResponse (Maybe AppInstanceUser)

The name of the AppInstanceUser.

describeAppInstanceUserResponse_httpStatus :: Lens' DescribeAppInstanceUserResponse Int

The response's http status code.

DescribeAppInstance

describeAppInstance_appInstanceArn :: Lens' DescribeAppInstance Text

The ARN of the AppInstance.

describeAppInstanceResponse_appInstance :: Lens' DescribeAppInstanceResponse (Maybe AppInstance)

The ARN, metadata, created and last-updated timestamps, and the name of the AppInstance. All timestamps use epoch milliseconds.

describeAppInstanceResponse_httpStatus :: Lens' DescribeAppInstanceResponse Int

The response's http status code.

ListAppInstanceUsers

listAppInstanceUsers_nextToken :: Lens' ListAppInstanceUsers (Maybe Text)

The token passed by previous API calls until all requested users are returned.

listAppInstanceUsers_maxResults :: Lens' ListAppInstanceUsers (Maybe Natural)

The maximum number of requests that you want returned.

listAppInstanceUsers_appInstanceArn :: Lens' ListAppInstanceUsers Text

The ARN of the AppInstance.

listAppInstanceUsersResponse_nextToken :: Lens' ListAppInstanceUsersResponse (Maybe Text)

The token passed by previous API calls until all requested users are returned.

listAppInstanceUsersResponse_appInstanceUsers :: Lens' ListAppInstanceUsersResponse (Maybe [AppInstanceUserSummary])

The information for each requested AppInstanceUser.

listAppInstanceUsersResponse_appInstanceArn :: Lens' ListAppInstanceUsersResponse (Maybe Text)

The ARN of the AppInstance.

listAppInstanceUsersResponse_httpStatus :: Lens' ListAppInstanceUsersResponse Int

The response's http status code.

DeleteAppInstanceUser

deleteAppInstanceUser_appInstanceUserArn :: Lens' DeleteAppInstanceUser Text

The ARN of the user request being deleted.

UpdateAppInstanceUser

updateAppInstanceUser_appInstanceUserArn :: Lens' UpdateAppInstanceUser Text

The ARN of the AppInstanceUser.

updateAppInstanceUser_name :: Lens' UpdateAppInstanceUser Text

The name of the AppInstanceUser.

updateAppInstanceUser_metadata :: Lens' UpdateAppInstanceUser Text

The metadata of the AppInstanceUser.

updateAppInstanceUserResponse_appInstanceUserArn :: Lens' UpdateAppInstanceUserResponse (Maybe Text)

The ARN of the AppInstanceUser.

updateAppInstanceUserResponse_httpStatus :: Lens' UpdateAppInstanceUserResponse Int

The response's http status code.

DeleteAppInstanceAdmin

deleteAppInstanceAdmin_appInstanceAdminArn :: Lens' DeleteAppInstanceAdmin Text

The ARN of the AppInstance's administrator.

deleteAppInstanceAdmin_appInstanceArn :: Lens' DeleteAppInstanceAdmin Text

The ARN of the AppInstance.

DeleteAppInstance

deleteAppInstance_appInstanceArn :: Lens' DeleteAppInstance Text

The ARN of the AppInstance.

UpdateAppInstance

updateAppInstance_appInstanceArn :: Lens' UpdateAppInstance Text

The ARN of the AppInstance.

updateAppInstance_name :: Lens' UpdateAppInstance Text

The name that you want to change.

updateAppInstance_metadata :: Lens' UpdateAppInstance Text

The metadata that you want to change.

updateAppInstanceResponse_appInstanceArn :: Lens' UpdateAppInstanceResponse (Maybe Text)

The ARN of the AppInstance.

updateAppInstanceResponse_httpStatus :: Lens' UpdateAppInstanceResponse Int

The response's http status code.

Types

AppInstance

appInstance_name :: Lens' AppInstance (Maybe Text)

The name of an AppInstance.

appInstance_metadata :: Lens' AppInstance (Maybe Text)

The metadata of an AppInstance.

appInstance_appInstanceArn :: Lens' AppInstance (Maybe Text)

The ARN of the messaging instance.

appInstance_createdTimestamp :: Lens' AppInstance (Maybe UTCTime)

The time at which an AppInstance was created. In epoch milliseconds.

appInstance_lastUpdatedTimestamp :: Lens' AppInstance (Maybe UTCTime)

The time an AppInstance was last updated. In epoch milliseconds.

AppInstanceAdmin

appInstanceAdmin_admin :: Lens' AppInstanceAdmin (Maybe Identity)

The AppInstanceAdmin data.

appInstanceAdmin_appInstanceArn :: Lens' AppInstanceAdmin (Maybe Text)

The ARN of the AppInstance for which the user is an administrator.

appInstanceAdmin_createdTimestamp :: Lens' AppInstanceAdmin (Maybe UTCTime)

The time at which an administrator was created.

AppInstanceAdminSummary

appInstanceAdminSummary_admin :: Lens' AppInstanceAdminSummary (Maybe Identity)

The details of the AppInstanceAdmin.

AppInstanceRetentionSettings

appInstanceRetentionSettings_channelRetentionSettings :: Lens' AppInstanceRetentionSettings (Maybe ChannelRetentionSettings)

The length of time in days to retain the messages in a channel.

AppInstanceSummary

appInstanceSummary_name :: Lens' AppInstanceSummary (Maybe Text)

The name of the AppInstance.

appInstanceSummary_metadata :: Lens' AppInstanceSummary (Maybe Text)

The metadata of the AppInstance.

appInstanceSummary_appInstanceArn :: Lens' AppInstanceSummary (Maybe Text)

The AppInstance ARN.

AppInstanceUser

appInstanceUser_appInstanceUserArn :: Lens' AppInstanceUser (Maybe Text)

The ARN of the AppInstanceUser.

appInstanceUser_name :: Lens' AppInstanceUser (Maybe Text)

The name of the AppInstanceUser.

appInstanceUser_metadata :: Lens' AppInstanceUser (Maybe Text)

The metadata of the AppInstanceUser.

appInstanceUser_createdTimestamp :: Lens' AppInstanceUser (Maybe UTCTime)

The time at which the AppInstanceUser was created.

appInstanceUser_lastUpdatedTimestamp :: Lens' AppInstanceUser (Maybe UTCTime)

The time at which the AppInstanceUser was last updated.

AppInstanceUserSummary

appInstanceUserSummary_appInstanceUserArn :: Lens' AppInstanceUserSummary (Maybe Text)

The ARN of the AppInstanceUser.

appInstanceUserSummary_name :: Lens' AppInstanceUserSummary (Maybe Text)

The name of an AppInstanceUser.

appInstanceUserSummary_metadata :: Lens' AppInstanceUserSummary (Maybe Text)

The metadata of the AppInstanceUser.

ChannelRetentionSettings

channelRetentionSettings_retentionDays :: Lens' ChannelRetentionSettings (Maybe Natural)

The time in days to retain the messages in a channel.

Identity

identity_arn :: Lens' Identity (Maybe Text)

The ARN in an Identity.

identity_name :: Lens' Identity (Maybe Text)

The name in an Identity.

Tag

tag_key :: Lens' Tag Text

The key of the tag.

tag_value :: Lens' Tag Text

The value of the tag.