Amazonka.ChimeSDKIdentity.ListAppInstanceUsers

Description

List all AppInstanceUsers created under a single AppInstance.

Creating a Request

data ListAppInstanceUsers

See: newListAppInstanceUsers smart constructor.

Constructors

ListAppInstanceUsers'

Fields

nextToken :: Maybe (Sensitive Text)

The token passed by previous API calls until all requested users are returned.

maxResults :: Maybe Natural

The maximum number of requests that you want returned.

appInstanceArn :: Text

The ARN of the AppInstance.

newListAppInstanceUsers

Arguments

:: Text
$sel:appInstanceArn:ListAppInstanceUsers'

-> ListAppInstanceUsers

Create a value of ListAppInstanceUsers with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:nextToken:ListAppInstanceUsers', listAppInstanceUsers_nextToken - The token passed by previous API calls until all requested users are returned.

$sel:maxResults:ListAppInstanceUsers', listAppInstanceUsers_maxResults - The maximum number of requests that you want returned.

$sel:appInstanceArn:ListAppInstanceUsers', listAppInstanceUsers_appInstanceArn - The ARN of the AppInstance.

Request Lenses

listAppInstanceUsers_nextToken :: Lens' ListAppInstanceUsers (Maybe Text)

The token passed by previous API calls until all requested users are returned.

listAppInstanceUsers_maxResults :: Lens' ListAppInstanceUsers (Maybe Natural)

The maximum number of requests that you want returned.

listAppInstanceUsers_appInstanceArn :: Lens' ListAppInstanceUsers Text

The ARN of the AppInstance.

Destructuring the Response

data ListAppInstanceUsersResponse

See: newListAppInstanceUsersResponse smart constructor.

Constructors

ListAppInstanceUsersResponse'

Fields

nextToken :: Maybe (Sensitive Text)

The token passed by previous API calls until all requested users are returned.

appInstanceUsers :: Maybe [AppInstanceUserSummary]

The information for each requested AppInstanceUser.

appInstanceArn :: Maybe Text

The ARN of the AppInstance.

httpStatus :: Int

The response's http status code.

newListAppInstanceUsersResponse

Arguments

:: Int
$sel:httpStatus:ListAppInstanceUsersResponse'

-> ListAppInstanceUsersResponse

Create a value of ListAppInstanceUsersResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:nextToken:ListAppInstanceUsers', listAppInstanceUsersResponse_nextToken - The token passed by previous API calls until all requested users are returned.

$sel:appInstanceUsers:ListAppInstanceUsersResponse', listAppInstanceUsersResponse_appInstanceUsers - The information for each requested AppInstanceUser.

$sel:appInstanceArn:ListAppInstanceUsers', listAppInstanceUsersResponse_appInstanceArn - The ARN of the AppInstance.

$sel:httpStatus:ListAppInstanceUsersResponse', listAppInstanceUsersResponse_httpStatus - The response's http status code.

Response Lenses

listAppInstanceUsersResponse_nextToken :: Lens' ListAppInstanceUsersResponse (Maybe Text)

The token passed by previous API calls until all requested users are returned.

listAppInstanceUsersResponse_appInstanceUsers :: Lens' ListAppInstanceUsersResponse (Maybe [AppInstanceUserSummary])

The information for each requested AppInstanceUser.

listAppInstanceUsersResponse_appInstanceArn :: Lens' ListAppInstanceUsersResponse (Maybe Text)

The ARN of the AppInstance.

listAppInstanceUsersResponse_httpStatus :: Lens' ListAppInstanceUsersResponse Int

The response's http status code.