Amazonka.ChimeSDKIdentity.ListAppInstances

Description

Lists all Amazon Chime AppInstances created under a single AWS account.

Creating a Request

data ListAppInstances

See: newListAppInstances smart constructor.

Constructors

ListAppInstances'

Fields

nextToken :: Maybe (Sensitive Text)

The token passed by previous API requests until you reach the maximum number of AppInstances.

maxResults :: Maybe Natural

The maximum number of AppInstances that you want to return.

newListAppInstances :: ListAppInstances

Create a value of ListAppInstances with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:nextToken:ListAppInstances', listAppInstances_nextToken - The token passed by previous API requests until you reach the maximum number of AppInstances.

$sel:maxResults:ListAppInstances', listAppInstances_maxResults - The maximum number of AppInstances that you want to return.

Request Lenses

listAppInstances_nextToken :: Lens' ListAppInstances (Maybe Text)

The token passed by previous API requests until you reach the maximum number of AppInstances.

listAppInstances_maxResults :: Lens' ListAppInstances (Maybe Natural)

The maximum number of AppInstances that you want to return.

Destructuring the Response

data ListAppInstancesResponse

See: newListAppInstancesResponse smart constructor.

Constructors

ListAppInstancesResponse'

Fields

appInstances :: Maybe [AppInstanceSummary]

The information for each AppInstance.

nextToken :: Maybe (Sensitive Text)

The token passed by previous API requests until the maximum number of AppInstances is reached.

httpStatus :: Int

The response's http status code.

newListAppInstancesResponse

Arguments

:: Int
$sel:httpStatus:ListAppInstancesResponse'

-> ListAppInstancesResponse

Create a value of ListAppInstancesResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:appInstances:ListAppInstancesResponse', listAppInstancesResponse_appInstances - The information for each AppInstance.

$sel:nextToken:ListAppInstances', listAppInstancesResponse_nextToken - The token passed by previous API requests until the maximum number of AppInstances is reached.

$sel:httpStatus:ListAppInstancesResponse', listAppInstancesResponse_httpStatus - The response's http status code.

Response Lenses

listAppInstancesResponse_appInstances :: Lens' ListAppInstancesResponse (Maybe [AppInstanceSummary])

The information for each AppInstance.

listAppInstancesResponse_nextToken :: Lens' ListAppInstancesResponse (Maybe Text)

The token passed by previous API requests until the maximum number of AppInstances is reached.

listAppInstancesResponse_httpStatus :: Lens' ListAppInstancesResponse Int

The response's http status code.