Amazonka.ChimeSDKMessaging.DescribeChannelBan

Description

Returns the full details of a channel ban.

The x-amz-chime-bearer request header is mandatory. Use the AppInstanceUserArn of the user that makes the API call as the value in the header.

Creating a Request

data DescribeChannelBan

See: newDescribeChannelBan smart constructor.

Constructors

DescribeChannelBan'

Fields

channelArn :: Text

The ARN of the channel from which the user is banned.

memberArn :: Text

The ARN of the member being banned.

chimeBearer :: Text

The AppInstanceUserArn of the user that makes the API call.

newDescribeChannelBan

Arguments

:: Text
$sel:channelArn:DescribeChannelBan'

-> Text
$sel:memberArn:DescribeChannelBan'

-> Text
$sel:chimeBearer:DescribeChannelBan'

-> DescribeChannelBan

Create a value of DescribeChannelBan with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:channelArn:DescribeChannelBan', describeChannelBan_channelArn - The ARN of the channel from which the user is banned.

$sel:memberArn:DescribeChannelBan', describeChannelBan_memberArn - The ARN of the member being banned.

$sel:chimeBearer:DescribeChannelBan', describeChannelBan_chimeBearer - The AppInstanceUserArn of the user that makes the API call.

Request Lenses

describeChannelBan_channelArn :: Lens' DescribeChannelBan Text

The ARN of the channel from which the user is banned.

describeChannelBan_memberArn :: Lens' DescribeChannelBan Text

The ARN of the member being banned.

describeChannelBan_chimeBearer :: Lens' DescribeChannelBan Text

The AppInstanceUserArn of the user that makes the API call.

Destructuring the Response

data DescribeChannelBanResponse

See: newDescribeChannelBanResponse smart constructor.

Constructors

DescribeChannelBanResponse'

Fields

channelBan :: Maybe ChannelBan

The details of the ban.

httpStatus :: Int

The response's http status code.

newDescribeChannelBanResponse

Arguments

:: Int
$sel:httpStatus:DescribeChannelBanResponse'

-> DescribeChannelBanResponse

Create a value of DescribeChannelBanResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:channelBan:DescribeChannelBanResponse', describeChannelBanResponse_channelBan - The details of the ban.

$sel:httpStatus:DescribeChannelBanResponse', describeChannelBanResponse_httpStatus - The response's http status code.

Response Lenses

describeChannelBanResponse_channelBan :: Lens' DescribeChannelBanResponse (Maybe ChannelBan)

The details of the ban.

describeChannelBanResponse_httpStatus :: Lens' DescribeChannelBanResponse Int

The response's http status code.