Amazonka.ChimeSDKMessaging.DescribeChannelMembershipForAppInstanceUser

Description

Returns the details of a channel based on the membership of the specified AppInstanceUser.

The x-amz-chime-bearer request header is mandatory. Use the AppInstanceUserArn of the user that makes the API call as the value in the header.

Creating a Request

data DescribeChannelMembershipForAppInstanceUser

See: newDescribeChannelMembershipForAppInstanceUser smart constructor.

Constructors

DescribeChannelMembershipForAppInstanceUser'

Fields

channelArn :: Text

The ARN of the channel to which the user belongs.

appInstanceUserArn :: Text

The ARN of the user in a channel.

chimeBearer :: Text

The AppInstanceUserArn of the user that makes the API call.

newDescribeChannelMembershipForAppInstanceUser

Arguments

:: Text
$sel:channelArn:DescribeChannelMembershipForAppInstanceUser'

-> Text
$sel:appInstanceUserArn:DescribeChannelMembershipForAppInstanceUser'

-> Text
$sel:chimeBearer:DescribeChannelMembershipForAppInstanceUser'

-> DescribeChannelMembershipForAppInstanceUser

Create a value of DescribeChannelMembershipForAppInstanceUser with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:channelArn:DescribeChannelMembershipForAppInstanceUser', describeChannelMembershipForAppInstanceUser_channelArn - The ARN of the channel to which the user belongs.

$sel:appInstanceUserArn:DescribeChannelMembershipForAppInstanceUser', describeChannelMembershipForAppInstanceUser_appInstanceUserArn - The ARN of the user in a channel.

$sel:chimeBearer:DescribeChannelMembershipForAppInstanceUser', describeChannelMembershipForAppInstanceUser_chimeBearer - The AppInstanceUserArn of the user that makes the API call.

Request Lenses

describeChannelMembershipForAppInstanceUser_channelArn :: Lens' DescribeChannelMembershipForAppInstanceUser Text

The ARN of the channel to which the user belongs.

describeChannelMembershipForAppInstanceUser_appInstanceUserArn :: Lens' DescribeChannelMembershipForAppInstanceUser Text

The ARN of the user in a channel.

describeChannelMembershipForAppInstanceUser_chimeBearer :: Lens' DescribeChannelMembershipForAppInstanceUser Text

The AppInstanceUserArn of the user that makes the API call.

Destructuring the Response

data DescribeChannelMembershipForAppInstanceUserResponse

See: newDescribeChannelMembershipForAppInstanceUserResponse smart constructor.

Constructors

DescribeChannelMembershipForAppInstanceUserResponse'

Fields

channelMembership :: Maybe ChannelMembershipForAppInstanceUserSummary

The channel to which a user belongs.

httpStatus :: Int

The response's http status code.

newDescribeChannelMembershipForAppInstanceUserResponse

Arguments

:: Int
$sel:httpStatus:DescribeChannelMembershipForAppInstanceUserResponse'

-> DescribeChannelMembershipForAppInstanceUserResponse

Create a value of DescribeChannelMembershipForAppInstanceUserResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:channelMembership:DescribeChannelMembershipForAppInstanceUserResponse', describeChannelMembershipForAppInstanceUserResponse_channelMembership - The channel to which a user belongs.

$sel:httpStatus:DescribeChannelMembershipForAppInstanceUserResponse', describeChannelMembershipForAppInstanceUserResponse_httpStatus - The response's http status code.

Response Lenses

describeChannelMembershipForAppInstanceUserResponse_channelMembership :: Lens' DescribeChannelMembershipForAppInstanceUserResponse (Maybe ChannelMembershipForAppInstanceUserSummary)

The channel to which a user belongs.

describeChannelMembershipForAppInstanceUserResponse_httpStatus :: Lens' DescribeChannelMembershipForAppInstanceUserResponse Int

The response's http status code.