Amazonka.ChimeSDKMessaging.DescribeChannelModerator

Description

Returns the full details of a single ChannelModerator.

The x-amz-chime-bearer request header is mandatory. Use the AppInstanceUserArn of the user that makes the API call as the value in the header.

Creating a Request

data DescribeChannelModerator

See: newDescribeChannelModerator smart constructor.

Constructors

DescribeChannelModerator'

Fields

channelArn :: Text

The ARN of the channel.

channelModeratorArn :: Text

The ARN of the channel moderator.

chimeBearer :: Text

The AppInstanceUserArn of the user that makes the API call.

newDescribeChannelModerator

Arguments

:: Text
$sel:channelArn:DescribeChannelModerator'

-> Text
$sel:channelModeratorArn:DescribeChannelModerator'

-> Text
$sel:chimeBearer:DescribeChannelModerator'

-> DescribeChannelModerator

Create a value of DescribeChannelModerator with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:channelArn:DescribeChannelModerator', describeChannelModerator_channelArn - The ARN of the channel.

$sel:channelModeratorArn:DescribeChannelModerator', describeChannelModerator_channelModeratorArn - The ARN of the channel moderator.

$sel:chimeBearer:DescribeChannelModerator', describeChannelModerator_chimeBearer - The AppInstanceUserArn of the user that makes the API call.

Request Lenses

describeChannelModerator_channelArn :: Lens' DescribeChannelModerator Text

The ARN of the channel.

describeChannelModerator_channelModeratorArn :: Lens' DescribeChannelModerator Text

The ARN of the channel moderator.

describeChannelModerator_chimeBearer :: Lens' DescribeChannelModerator Text

The AppInstanceUserArn of the user that makes the API call.

Destructuring the Response

data DescribeChannelModeratorResponse

See: newDescribeChannelModeratorResponse smart constructor.

Constructors

DescribeChannelModeratorResponse'

Fields

channelModerator :: Maybe ChannelModerator

The details of the channel moderator.

httpStatus :: Int

The response's http status code.

newDescribeChannelModeratorResponse

Arguments

:: Int
$sel:httpStatus:DescribeChannelModeratorResponse'

-> DescribeChannelModeratorResponse

Create a value of DescribeChannelModeratorResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:channelModerator:DescribeChannelModeratorResponse', describeChannelModeratorResponse_channelModerator - The details of the channel moderator.

$sel:httpStatus:DescribeChannelModeratorResponse', describeChannelModeratorResponse_httpStatus - The response's http status code.

Response Lenses

describeChannelModeratorResponse_channelModerator :: Lens' DescribeChannelModeratorResponse (Maybe ChannelModerator)

The details of the channel moderator.

describeChannelModeratorResponse_httpStatus :: Lens' DescribeChannelModeratorResponse Int

The response's http status code.