Amazonka.ChimeSDKMessaging.Lens

Description

Operations

DescribeChannelMembership

describeChannelMembership_channelArn :: Lens' DescribeChannelMembership Text

The ARN of the channel.

describeChannelMembership_memberArn :: Lens' DescribeChannelMembership Text

The ARN of the member.

describeChannelMembership_chimeBearer :: Lens' DescribeChannelMembership Text

The AppInstanceUserArn of the user that makes the API call.

describeChannelMembershipResponse_channelMembership :: Lens' DescribeChannelMembershipResponse (Maybe ChannelMembership)

The details of the membership.

describeChannelMembershipResponse_httpStatus :: Lens' DescribeChannelMembershipResponse Int

The response's http status code.

DescribeChannelFlow

describeChannelFlow_channelFlowArn :: Lens' DescribeChannelFlow Text

The ARN of the channel flow.

describeChannelFlowResponse_channelFlow :: Lens' DescribeChannelFlowResponse (Maybe ChannelFlow)

The channel flow details.

describeChannelFlowResponse_httpStatus :: Lens' DescribeChannelFlowResponse Int

The response's http status code.

ListChannels

listChannels_privacy :: Lens' ListChannels (Maybe ChannelPrivacy)

The privacy setting. PUBLIC retrieves all the public channels. PRIVATE retrieves private channels. Only an AppInstanceAdmin can retrieve private channels.

listChannels_nextToken :: Lens' ListChannels (Maybe Text)

The token passed by previous API calls until all requested channels are returned.

listChannels_maxResults :: Lens' ListChannels (Maybe Natural)

The maximum number of channels that you want to return.

listChannels_appInstanceArn :: Lens' ListChannels Text

The ARN of the AppInstance.

listChannels_chimeBearer :: Lens' ListChannels Text

The AppInstanceUserArn of the user that makes the API call.

listChannelsResponse_channels :: Lens' ListChannelsResponse (Maybe [ChannelSummary])

The information about each channel.

listChannelsResponse_nextToken :: Lens' ListChannelsResponse (Maybe Text)

The token returned from previous API requests until the number of channels is reached.

listChannelsResponse_httpStatus :: Lens' ListChannelsResponse Int

The response's http status code.

ListTagsForResource

listTagsForResource_resourceARN :: Lens' ListTagsForResource Text

The ARN of the resource.

listTagsForResourceResponse_tags :: Lens' ListTagsForResourceResponse (Maybe (NonEmpty Tag))

The tag key-value pairs.

listTagsForResourceResponse_httpStatus :: Lens' ListTagsForResourceResponse Int

The response's http status code.

DeleteChannel

deleteChannel_channelArn :: Lens' DeleteChannel Text

The ARN of the channel being deleted.

deleteChannel_chimeBearer :: Lens' DeleteChannel Text

The AppInstanceUserArn of the user that makes the API call.

UpdateChannel

updateChannel_metadata :: Lens' UpdateChannel (Maybe Text)

The metadata for the update request.

updateChannel_channelArn :: Lens' UpdateChannel Text

The ARN of the channel.

updateChannel_name :: Lens' UpdateChannel Text

The name of the channel.

updateChannel_mode :: Lens' UpdateChannel ChannelMode

The mode of the update request.

updateChannel_chimeBearer :: Lens' UpdateChannel Text

The AppInstanceUserArn of the user that makes the API call.

updateChannelResponse_channelArn :: Lens' UpdateChannelResponse (Maybe Text)

The ARN of the channel.

updateChannelResponse_httpStatus :: Lens' UpdateChannelResponse Int

The response's http status code.

AssociateChannelFlow

associateChannelFlow_channelArn :: Lens' AssociateChannelFlow Text

The ARN of the channel.

associateChannelFlow_channelFlowArn :: Lens' AssociateChannelFlow Text

The ARN of the channel flow.

associateChannelFlow_chimeBearer :: Lens' AssociateChannelFlow Text

The AppInstanceUserArn of the user making the API call.

GetMessagingSessionEndpoint

getMessagingSessionEndpointResponse_endpoint :: Lens' GetMessagingSessionEndpointResponse (Maybe MessagingSessionEndpoint)

The endpoint returned in the response.

getMessagingSessionEndpointResponse_httpStatus :: Lens' GetMessagingSessionEndpointResponse Int

The response's http status code.

ListChannelsModeratedByAppInstanceUser

listChannelsModeratedByAppInstanceUser_appInstanceUserArn :: Lens' ListChannelsModeratedByAppInstanceUser (Maybe Text)

The ARN of the user in the moderated channel.

listChannelsModeratedByAppInstanceUser_nextToken :: Lens' ListChannelsModeratedByAppInstanceUser (Maybe Text)

The token returned from previous API requests until the number of channels moderated by the user is reached.

listChannelsModeratedByAppInstanceUser_maxResults :: Lens' ListChannelsModeratedByAppInstanceUser (Maybe Natural)

The maximum number of channels in the request.

listChannelsModeratedByAppInstanceUser_chimeBearer :: Lens' ListChannelsModeratedByAppInstanceUser Text

The AppInstanceUserArn of the user that makes the API call.

listChannelsModeratedByAppInstanceUserResponse_channels :: Lens' ListChannelsModeratedByAppInstanceUserResponse (Maybe [ChannelModeratedByAppInstanceUserSummary])

The moderated channels in the request.

listChannelsModeratedByAppInstanceUserResponse_nextToken :: Lens' ListChannelsModeratedByAppInstanceUserResponse (Maybe Text)

The token returned from previous API requests until the number of channels moderated by the user is reached.

listChannelsModeratedByAppInstanceUserResponse_httpStatus :: Lens' ListChannelsModeratedByAppInstanceUserResponse Int

The response's http status code.

RedactChannelMessage

redactChannelMessage_channelArn :: Lens' RedactChannelMessage Text

The ARN of the channel containing the messages that you want to redact.

redactChannelMessage_messageId :: Lens' RedactChannelMessage Text

The ID of the message being redacted.

redactChannelMessage_chimeBearer :: Lens' RedactChannelMessage Text

The AppInstanceUserArn of the user that makes the API call.

redactChannelMessageResponse_channelArn :: Lens' RedactChannelMessageResponse (Maybe Text)

The ARN of the channel containing the messages that you want to redact.

redactChannelMessageResponse_messageId :: Lens' RedactChannelMessageResponse (Maybe Text)

The ID of the message being redacted.

redactChannelMessageResponse_httpStatus :: Lens' RedactChannelMessageResponse Int

The response's http status code.

ListChannelFlows

listChannelFlows_nextToken :: Lens' ListChannelFlows (Maybe Text)

The token passed by previous API calls until all requested channel flows are returned.

listChannelFlows_maxResults :: Lens' ListChannelFlows (Maybe Natural)

The maximum number of channel flows that you want to return.

listChannelFlows_appInstanceArn :: Lens' ListChannelFlows Text

The ARN of the app instance.

listChannelFlowsResponse_channelFlows :: Lens' ListChannelFlowsResponse (Maybe [ChannelFlowSummary])

The information about each channel flow.

listChannelFlowsResponse_nextToken :: Lens' ListChannelFlowsResponse (Maybe Text)

The token passed by previous API calls until all requested channels are returned.

listChannelFlowsResponse_httpStatus :: Lens' ListChannelFlowsResponse Int

The response's http status code.

DeleteChannelFlow

deleteChannelFlow_channelFlowArn :: Lens' DeleteChannelFlow Text

The ARN of the channel flow.

UpdateChannelFlow

updateChannelFlow_channelFlowArn :: Lens' UpdateChannelFlow Text

The ARN of the channel flow.

updateChannelFlow_processors :: Lens' UpdateChannelFlow (NonEmpty Processor)

Information about the processor Lambda functions

updateChannelFlow_name :: Lens' UpdateChannelFlow Text

The name of the channel flow.

updateChannelFlowResponse_channelFlowArn :: Lens' UpdateChannelFlowResponse (Maybe Text)

The ARN of the channel flow.

updateChannelFlowResponse_httpStatus :: Lens' UpdateChannelFlowResponse Int

The response's http status code.

DeleteChannelMembership

deleteChannelMembership_channelArn :: Lens' DeleteChannelMembership Text

The ARN of the channel from which you want to remove the user.

deleteChannelMembership_memberArn :: Lens' DeleteChannelMembership Text

The ARN of the member that you're removing from the channel.

deleteChannelMembership_chimeBearer :: Lens' DeleteChannelMembership Text

The AppInstanceUserArn of the user that makes the API call.

ListChannelMemberships

listChannelMemberships_nextToken :: Lens' ListChannelMemberships (Maybe Text)

The token passed by previous API calls until all requested channel memberships are returned.

listChannelMemberships_type :: Lens' ListChannelMemberships (Maybe ChannelMembershipType)

The membership type of a user, DEFAULT or HIDDEN. Default members are always returned as part of ListChannelMemberships. Hidden members are only returned if the type filter in ListChannelMemberships equals HIDDEN. Otherwise hidden members are not returned.

listChannelMemberships_maxResults :: Lens' ListChannelMemberships (Maybe Natural)

The maximum number of channel memberships that you want returned.

listChannelMemberships_channelArn :: Lens' ListChannelMemberships Text

The maximum number of channel memberships that you want returned.

listChannelMemberships_chimeBearer :: Lens' ListChannelMemberships Text

The AppInstanceUserArn of the user that makes the API call.

listChannelMembershipsResponse_channelMemberships :: Lens' ListChannelMembershipsResponse (Maybe [ChannelMembershipSummary])

The information for the requested channel memberships.

listChannelMembershipsResponse_channelArn :: Lens' ListChannelMembershipsResponse (Maybe Text)

The ARN of the channel.

listChannelMembershipsResponse_nextToken :: Lens' ListChannelMembershipsResponse (Maybe Text)

The token passed by previous API calls until all requested channel memberships are returned.

listChannelMembershipsResponse_httpStatus :: Lens' ListChannelMembershipsResponse Int

The response's http status code.

DisassociateChannelFlow

disassociateChannelFlow_channelArn :: Lens' DisassociateChannelFlow Text

The ARN of the channel.

disassociateChannelFlow_channelFlowArn :: Lens' DisassociateChannelFlow Text

The ARN of the channel flow.

disassociateChannelFlow_chimeBearer :: Lens' DisassociateChannelFlow Text

The AppInstanceUserArn of the user making the API call.

GetChannelMessage

getChannelMessage_channelArn :: Lens' GetChannelMessage Text

The ARN of the channel.

getChannelMessage_messageId :: Lens' GetChannelMessage Text

The ID of the message.

getChannelMessage_chimeBearer :: Lens' GetChannelMessage Text

The AppInstanceUserArn of the user that makes the API call.

getChannelMessageResponse_channelMessage :: Lens' GetChannelMessageResponse (Maybe ChannelMessage)

The details of and content in the message.

getChannelMessageResponse_httpStatus :: Lens' GetChannelMessageResponse Int

The response's http status code.

DescribeChannelMembershipForAppInstanceUser

describeChannelMembershipForAppInstanceUser_channelArn :: Lens' DescribeChannelMembershipForAppInstanceUser Text

The ARN of the channel to which the user belongs.

describeChannelMembershipForAppInstanceUser_appInstanceUserArn :: Lens' DescribeChannelMembershipForAppInstanceUser Text

The ARN of the user in a channel.

describeChannelMembershipForAppInstanceUser_chimeBearer :: Lens' DescribeChannelMembershipForAppInstanceUser Text

The AppInstanceUserArn of the user that makes the API call.

describeChannelMembershipForAppInstanceUserResponse_channelMembership :: Lens' DescribeChannelMembershipForAppInstanceUserResponse (Maybe ChannelMembershipForAppInstanceUserSummary)

The channel to which a user belongs.

describeChannelMembershipForAppInstanceUserResponse_httpStatus :: Lens' DescribeChannelMembershipForAppInstanceUserResponse Int

The response's http status code.

CreateChannelModerator

createChannelModerator_channelArn :: Lens' CreateChannelModerator Text

The ARN of the channel.

createChannelModerator_channelModeratorArn :: Lens' CreateChannelModerator Text

The ARN of the moderator.

createChannelModerator_chimeBearer :: Lens' CreateChannelModerator Text

The AppInstanceUserArn of the user that makes the API call.

createChannelModeratorResponse_channelModerator :: Lens' CreateChannelModeratorResponse (Maybe Identity)

The ARNs of the channel and the moderator.

createChannelModeratorResponse_channelArn :: Lens' CreateChannelModeratorResponse (Maybe Text)

The ARN of the channel.

createChannelModeratorResponse_httpStatus :: Lens' CreateChannelModeratorResponse Int

The response's http status code.

DescribeChannelModeratedByAppInstanceUser

describeChannelModeratedByAppInstanceUser_channelArn :: Lens' DescribeChannelModeratedByAppInstanceUser Text

The ARN of the moderated channel.

describeChannelModeratedByAppInstanceUser_appInstanceUserArn :: Lens' DescribeChannelModeratedByAppInstanceUser Text

The ARN of the AppInstanceUser in the moderated channel.

describeChannelModeratedByAppInstanceUser_chimeBearer :: Lens' DescribeChannelModeratedByAppInstanceUser Text

The AppInstanceUserArn of the user that makes the API call.

describeChannelModeratedByAppInstanceUserResponse_channel :: Lens' DescribeChannelModeratedByAppInstanceUserResponse (Maybe ChannelModeratedByAppInstanceUserSummary)

The moderated channel.

describeChannelModeratedByAppInstanceUserResponse_httpStatus :: Lens' DescribeChannelModeratedByAppInstanceUserResponse Int

The response's http status code.

SendChannelMessage

sendChannelMessage_metadata :: Lens' SendChannelMessage (Maybe Text)

The optional metadata for each message.

sendChannelMessage_channelArn :: Lens' SendChannelMessage Text

The ARN of the channel.

sendChannelMessage_content :: Lens' SendChannelMessage Text

The content of the message.

sendChannelMessage_type :: Lens' SendChannelMessage ChannelMessageType

The type of message, STANDARD or CONTROL.

sendChannelMessage_persistence :: Lens' SendChannelMessage ChannelMessagePersistenceType

Boolean that controls whether the message is persisted on the back end. Required.

sendChannelMessage_clientRequestToken :: Lens' SendChannelMessage Text

The Idempotency token for each client request.

sendChannelMessage_chimeBearer :: Lens' SendChannelMessage Text

The AppInstanceUserArn of the user that makes the API call.

sendChannelMessageResponse_status :: Lens' SendChannelMessageResponse (Maybe ChannelMessageStatusStructure)

The status of the channel message.

sendChannelMessageResponse_channelArn :: Lens' SendChannelMessageResponse (Maybe Text)

The ARN of the channel.

sendChannelMessageResponse_messageId :: Lens' SendChannelMessageResponse (Maybe Text)

The ID string assigned to each message.

sendChannelMessageResponse_httpStatus :: Lens' SendChannelMessageResponse Int

The response's http status code.

DeleteChannelBan

deleteChannelBan_channelArn :: Lens' DeleteChannelBan Text

The ARN of the channel from which the AppInstanceUser was banned.

deleteChannelBan_memberArn :: Lens' DeleteChannelBan Text

The ARN of the AppInstanceUser that you want to reinstate.

deleteChannelBan_chimeBearer :: Lens' DeleteChannelBan Text

The AppInstanceUserArn of the user that makes the API call.

ListChannelBans

listChannelBans_nextToken :: Lens' ListChannelBans (Maybe Text)

The token passed by previous API calls until all requested bans are returned.

listChannelBans_maxResults :: Lens' ListChannelBans (Maybe Natural)

The maximum number of bans that you want returned.

listChannelBans_channelArn :: Lens' ListChannelBans Text

The ARN of the channel.

listChannelBans_chimeBearer :: Lens' ListChannelBans Text

The AppInstanceUserArn of the user that makes the API call.

listChannelBansResponse_channelArn :: Lens' ListChannelBansResponse (Maybe Text)

The ARN of the channel.

listChannelBansResponse_nextToken :: Lens' ListChannelBansResponse (Maybe Text)

The token passed by previous API calls until all requested bans are returned.

listChannelBansResponse_channelBans :: Lens' ListChannelBansResponse (Maybe [ChannelBanSummary])

The information for each requested ban.

listChannelBansResponse_httpStatus :: Lens' ListChannelBansResponse Int

The response's http status code.

CreateChannel

createChannel_mode :: Lens' CreateChannel (Maybe ChannelMode)

The channel mode: UNRESTRICTED or RESTRICTED. Administrators, moderators, and channel members can add themselves and other members to unrestricted channels. Only administrators and moderators can add members to restricted channels.

createChannel_privacy :: Lens' CreateChannel (Maybe ChannelPrivacy)

The channel's privacy level: PUBLIC or PRIVATE. Private channels aren't discoverable by users outside the channel. Public channels are discoverable by anyone in the AppInstance.

createChannel_metadata :: Lens' CreateChannel (Maybe Text)

The metadata of the creation request. Limited to 1KB and UTF-8.

createChannel_tags :: Lens' CreateChannel (Maybe (NonEmpty Tag))

The tags for the creation request.

createChannel_appInstanceArn :: Lens' CreateChannel Text

The ARN of the channel request.

createChannel_name :: Lens' CreateChannel Text

The name of the channel.

createChannel_clientRequestToken :: Lens' CreateChannel Text

The client token for the request. An Idempotency token.

createChannel_chimeBearer :: Lens' CreateChannel Text

The AppInstanceUserArn of the user that makes the API call.

createChannelResponse_channelArn :: Lens' CreateChannelResponse (Maybe Text)

The ARN of the channel.

createChannelResponse_httpStatus :: Lens' CreateChannelResponse Int

The response's http status code.

DescribeChannelModerator

describeChannelModerator_channelArn :: Lens' DescribeChannelModerator Text

The ARN of the channel.

describeChannelModerator_channelModeratorArn :: Lens' DescribeChannelModerator Text

The ARN of the channel moderator.

describeChannelModerator_chimeBearer :: Lens' DescribeChannelModerator Text

The AppInstanceUserArn of the user that makes the API call.

describeChannelModeratorResponse_channelModerator :: Lens' DescribeChannelModeratorResponse (Maybe ChannelModerator)

The details of the channel moderator.

describeChannelModeratorResponse_httpStatus :: Lens' DescribeChannelModeratorResponse Int

The response's http status code.

CreateChannelBan

createChannelBan_channelArn :: Lens' CreateChannelBan Text

The ARN of the ban request.

createChannelBan_memberArn :: Lens' CreateChannelBan Text

The ARN of the member being banned.

createChannelBan_chimeBearer :: Lens' CreateChannelBan Text

The AppInstanceUserArn of the user that makes the API call.

createChannelBanResponse_channelArn :: Lens' CreateChannelBanResponse (Maybe Text)

The ARN of the response to the ban request.

createChannelBanResponse_member :: Lens' CreateChannelBanResponse (Maybe Identity)

The ChannelArn and BannedIdentity of the member in the ban response.

createChannelBanResponse_httpStatus :: Lens' CreateChannelBanResponse Int

The response's http status code.

ListChannelMembershipsForAppInstanceUser

listChannelMembershipsForAppInstanceUser_appInstanceUserArn :: Lens' ListChannelMembershipsForAppInstanceUser (Maybe Text)

The ARN of the AppInstanceUsers

listChannelMembershipsForAppInstanceUser_nextToken :: Lens' ListChannelMembershipsForAppInstanceUser (Maybe Text)

The token returned from previous API requests until the number of channel memberships is reached.

listChannelMembershipsForAppInstanceUser_maxResults :: Lens' ListChannelMembershipsForAppInstanceUser (Maybe Natural)

The maximum number of users that you want returned.

listChannelMembershipsForAppInstanceUser_chimeBearer :: Lens' ListChannelMembershipsForAppInstanceUser Text

The AppInstanceUserArn of the user that makes the API call.

listChannelMembershipsForAppInstanceUserResponse_channelMemberships :: Lens' ListChannelMembershipsForAppInstanceUserResponse (Maybe [ChannelMembershipForAppInstanceUserSummary])

The token passed by previous API calls until all requested users are returned.

listChannelMembershipsForAppInstanceUserResponse_nextToken :: Lens' ListChannelMembershipsForAppInstanceUserResponse (Maybe Text)

The token passed by previous API calls until all requested users are returned.

listChannelMembershipsForAppInstanceUserResponse_httpStatus :: Lens' ListChannelMembershipsForAppInstanceUserResponse Int

The response's http status code.

UpdateChannelReadMarker

updateChannelReadMarker_channelArn :: Lens' UpdateChannelReadMarker Text

The ARN of the channel.

updateChannelReadMarker_chimeBearer :: Lens' UpdateChannelReadMarker Text

The AppInstanceUserArn of the user that makes the API call.

updateChannelReadMarkerResponse_channelArn :: Lens' UpdateChannelReadMarkerResponse (Maybe Text)

The ARN of the channel.

updateChannelReadMarkerResponse_httpStatus :: Lens' UpdateChannelReadMarkerResponse Int

The response's http status code.

GetChannelMessageStatus

getChannelMessageStatus_channelArn :: Lens' GetChannelMessageStatus Text

The ARN of the channel

getChannelMessageStatus_messageId :: Lens' GetChannelMessageStatus Text

The ID of the message.

getChannelMessageStatus_chimeBearer :: Lens' GetChannelMessageStatus Text

The AppInstanceUserArn of the user making the API call.

getChannelMessageStatusResponse_status :: Lens' GetChannelMessageStatusResponse (Maybe ChannelMessageStatusStructure)

The message status and details.

getChannelMessageStatusResponse_httpStatus :: Lens' GetChannelMessageStatusResponse Int

The response's http status code.

CreateChannelFlow

createChannelFlow_tags :: Lens' CreateChannelFlow (Maybe (NonEmpty Tag))

The tags for the creation request.

createChannelFlow_appInstanceArn :: Lens' CreateChannelFlow Text

The ARN of the channel flow request.

createChannelFlow_processors :: Lens' CreateChannelFlow (NonEmpty Processor)

Information about the processor Lambda functions.

createChannelFlow_name :: Lens' CreateChannelFlow Text

The name of the channel flow.

createChannelFlow_clientRequestToken :: Lens' CreateChannelFlow Text

The client token for the request. An Idempotency token.

createChannelFlowResponse_channelFlowArn :: Lens' CreateChannelFlowResponse (Maybe Text)

The ARN of the channel flow.

createChannelFlowResponse_httpStatus :: Lens' CreateChannelFlowResponse Int

The response's http status code.

CreateChannelMembership

createChannelMembership_channelArn :: Lens' CreateChannelMembership Text

The ARN of the channel to which you're adding users.

createChannelMembership_memberArn :: Lens' CreateChannelMembership Text

The ARN of the member you want to add to the channel.

createChannelMembership_type :: Lens' CreateChannelMembership ChannelMembershipType

The membership type of a user, DEFAULT or HIDDEN. Default members are always returned as part of ListChannelMemberships. Hidden members are only returned if the type filter in ListChannelMemberships equals HIDDEN. Otherwise hidden members are not returned. This is only supported by moderators.

createChannelMembership_chimeBearer :: Lens' CreateChannelMembership Text

The AppInstanceUserArn of the user that makes the API call.

createChannelMembershipResponse_channelArn :: Lens' CreateChannelMembershipResponse (Maybe Text)

The ARN of the channel.

createChannelMembershipResponse_member :: Lens' CreateChannelMembershipResponse (Maybe Identity)

The ARN and metadata of the member being added.

createChannelMembershipResponse_httpStatus :: Lens' CreateChannelMembershipResponse Int

The response's http status code.

TagResource

tagResource_resourceARN :: Lens' TagResource Text

The resource ARN.

tagResource_tags :: Lens' TagResource (NonEmpty Tag)

The tag key-value pairs.

ChannelFlowCallback

channelFlowCallback_deleteResource :: Lens' ChannelFlowCallback (Maybe Bool)

When a processor determines that a message needs to be DENIED, pass this parameter with a value of true.

channelFlowCallback_callbackId :: Lens' ChannelFlowCallback Text

The identifier passed to the processor by the service when invoked. Use the identifier to call back the service.

channelFlowCallback_channelArn :: Lens' ChannelFlowCallback Text

The ARN of the channel.

channelFlowCallback_channelMessage :: Lens' ChannelFlowCallback ChannelMessageCallback

Stores information about the processed message.

channelFlowCallbackResponse_callbackId :: Lens' ChannelFlowCallbackResponse (Maybe Text)

The call back ID passed in the request.

channelFlowCallbackResponse_channelArn :: Lens' ChannelFlowCallbackResponse (Maybe Text)

The ARN of the channel.

channelFlowCallbackResponse_httpStatus :: Lens' ChannelFlowCallbackResponse Int

The response's http status code.

DeleteChannelModerator

deleteChannelModerator_channelArn :: Lens' DeleteChannelModerator Text

The ARN of the channel.

deleteChannelModerator_channelModeratorArn :: Lens' DeleteChannelModerator Text

The ARN of the moderator being deleted.

deleteChannelModerator_chimeBearer :: Lens' DeleteChannelModerator Text

The AppInstanceUserArn of the user that makes the API call.

DescribeChannelBan

describeChannelBan_channelArn :: Lens' DescribeChannelBan Text

The ARN of the channel from which the user is banned.

describeChannelBan_memberArn :: Lens' DescribeChannelBan Text

The ARN of the member being banned.

describeChannelBan_chimeBearer :: Lens' DescribeChannelBan Text

The AppInstanceUserArn of the user that makes the API call.

describeChannelBanResponse_channelBan :: Lens' DescribeChannelBanResponse (Maybe ChannelBan)

The details of the ban.

describeChannelBanResponse_httpStatus :: Lens' DescribeChannelBanResponse Int

The response's http status code.

UntagResource

untagResource_resourceARN :: Lens' UntagResource Text

The resource ARN.

untagResource_tagKeys :: Lens' UntagResource (NonEmpty Text)

The tag keys.

ListChannelModerators

listChannelModerators_nextToken :: Lens' ListChannelModerators (Maybe Text)

The token passed by previous API calls until all requested moderators are returned.

listChannelModerators_maxResults :: Lens' ListChannelModerators (Maybe Natural)

The maximum number of moderators that you want returned.

listChannelModerators_channelArn :: Lens' ListChannelModerators Text

The ARN of the channel.

listChannelModerators_chimeBearer :: Lens' ListChannelModerators Text

The AppInstanceUserArn of the user that makes the API call.

listChannelModeratorsResponse_channelArn :: Lens' ListChannelModeratorsResponse (Maybe Text)

The ARN of the channel.

listChannelModeratorsResponse_nextToken :: Lens' ListChannelModeratorsResponse (Maybe Text)

The token passed by previous API calls until all requested moderators are returned.

listChannelModeratorsResponse_channelModerators :: Lens' ListChannelModeratorsResponse (Maybe [ChannelModeratorSummary])

The information about and names of each moderator.

listChannelModeratorsResponse_httpStatus :: Lens' ListChannelModeratorsResponse Int

The response's http status code.

DescribeChannel

describeChannel_channelArn :: Lens' DescribeChannel Text

The ARN of the channel.

describeChannel_chimeBearer :: Lens' DescribeChannel Text

The AppInstanceUserArn of the user that makes the API call.

describeChannelResponse_channel :: Lens' DescribeChannelResponse (Maybe Channel)

The channel details.

describeChannelResponse_httpStatus :: Lens' DescribeChannelResponse Int

The response's http status code.

DeleteChannelMessage

deleteChannelMessage_channelArn :: Lens' DeleteChannelMessage Text

The ARN of the channel.

deleteChannelMessage_messageId :: Lens' DeleteChannelMessage Text

The ID of the message being deleted.

deleteChannelMessage_chimeBearer :: Lens' DeleteChannelMessage Text

The AppInstanceUserArn of the user that makes the API call.

UpdateChannelMessage

updateChannelMessage_content :: Lens' UpdateChannelMessage (Maybe Text)

The content of the message being updated.

updateChannelMessage_metadata :: Lens' UpdateChannelMessage (Maybe Text)

The metadata of the message being updated.

updateChannelMessage_channelArn :: Lens' UpdateChannelMessage Text

The ARN of the channel.

updateChannelMessage_messageId :: Lens' UpdateChannelMessage Text

The ID string of the message being updated.

updateChannelMessage_chimeBearer :: Lens' UpdateChannelMessage Text

The AppInstanceUserArn of the user that makes the API call.

updateChannelMessageResponse_status :: Lens' UpdateChannelMessageResponse (Maybe ChannelMessageStatusStructure)

The status of the message update.

updateChannelMessageResponse_channelArn :: Lens' UpdateChannelMessageResponse (Maybe Text)

The ARN of the channel.

updateChannelMessageResponse_messageId :: Lens' UpdateChannelMessageResponse (Maybe Text)

The ID string of the message being updated.

updateChannelMessageResponse_httpStatus :: Lens' UpdateChannelMessageResponse Int

The response's http status code.

ListChannelMessages

listChannelMessages_nextToken :: Lens' ListChannelMessages (Maybe Text)

The token passed by previous API calls until all requested messages are returned.

listChannelMessages_notBefore :: Lens' ListChannelMessages (Maybe UTCTime)

The initial or starting time stamp for your requested messages.

listChannelMessages_sortOrder :: Lens' ListChannelMessages (Maybe SortOrder)

The order in which you want messages sorted. Default is Descending, based on time created.

listChannelMessages_maxResults :: Lens' ListChannelMessages (Maybe Natural)

The maximum number of messages that you want returned.

listChannelMessages_notAfter :: Lens' ListChannelMessages (Maybe UTCTime)

The final or ending time stamp for your requested messages.

listChannelMessages_channelArn :: Lens' ListChannelMessages Text

The ARN of the channel.

listChannelMessages_chimeBearer :: Lens' ListChannelMessages Text

The AppInstanceUserArn of the user that makes the API call.

listChannelMessagesResponse_channelArn :: Lens' ListChannelMessagesResponse (Maybe Text)

The ARN of the channel containing the requested messages.

listChannelMessagesResponse_nextToken :: Lens' ListChannelMessagesResponse (Maybe Text)

The token passed by previous API calls until all requested messages are returned.

listChannelMessagesResponse_channelMessages :: Lens' ListChannelMessagesResponse (Maybe [ChannelMessageSummary])

The information about, and content of, each requested message.

listChannelMessagesResponse_httpStatus :: Lens' ListChannelMessagesResponse Int

The response's http status code.

ListChannelsAssociatedWithChannelFlow

listChannelsAssociatedWithChannelFlow_nextToken :: Lens' ListChannelsAssociatedWithChannelFlow (Maybe Text)

The token passed by previous API calls until all requested channels are returned.

listChannelsAssociatedWithChannelFlow_maxResults :: Lens' ListChannelsAssociatedWithChannelFlow (Maybe Natural)

The maximum number of channels that you want to return.

listChannelsAssociatedWithChannelFlow_channelFlowArn :: Lens' ListChannelsAssociatedWithChannelFlow Text

The ARN of the channel flow.

listChannelsAssociatedWithChannelFlowResponse_channels :: Lens' ListChannelsAssociatedWithChannelFlowResponse (Maybe [ChannelAssociatedWithFlowSummary])

The information about each channel.

listChannelsAssociatedWithChannelFlowResponse_nextToken :: Lens' ListChannelsAssociatedWithChannelFlowResponse (Maybe Text)

The token passed by previous API calls until all requested channels are returned.

listChannelsAssociatedWithChannelFlowResponse_httpStatus :: Lens' ListChannelsAssociatedWithChannelFlowResponse Int

The response's http status code.

BatchCreateChannelMembership

batchCreateChannelMembership_type :: Lens' BatchCreateChannelMembership (Maybe ChannelMembershipType)

The membership type of a user, DEFAULT or HIDDEN. Default members are always returned as part of ListChannelMemberships. Hidden members are only returned if the type filter in ListChannelMemberships equals HIDDEN. Otherwise hidden members are not returned. This is only supported by moderators.

batchCreateChannelMembership_channelArn :: Lens' BatchCreateChannelMembership Text

The ARN of the channel to which you're adding users.

batchCreateChannelMembership_memberArns :: Lens' BatchCreateChannelMembership (NonEmpty Text)

The ARNs of the members you want to add to the channel.

batchCreateChannelMembership_chimeBearer :: Lens' BatchCreateChannelMembership Text

The AppInstanceUserArn of the user that makes the API call.

batchCreateChannelMembershipResponse_errors :: Lens' BatchCreateChannelMembershipResponse (Maybe [BatchCreateChannelMembershipError])

If the action fails for one or more of the memberships in the request, a list of the memberships is returned, along with error codes and error messages.

batchCreateChannelMembershipResponse_batchChannelMemberships :: Lens' BatchCreateChannelMembershipResponse (Maybe BatchChannelMemberships)

The list of channel memberships in the response.

batchCreateChannelMembershipResponse_httpStatus :: Lens' BatchCreateChannelMembershipResponse Int

The response's http status code.

Types

AppInstanceUserMembershipSummary

appInstanceUserMembershipSummary_readMarkerTimestamp :: Lens' AppInstanceUserMembershipSummary (Maybe UTCTime)

The time at which a message was last read.

appInstanceUserMembershipSummary_type :: Lens' AppInstanceUserMembershipSummary (Maybe ChannelMembershipType)

The type of ChannelMembership.

BatchChannelMemberships

batchChannelMemberships_members :: Lens' BatchChannelMemberships (Maybe [Identity])

The users successfully added to the request.

batchChannelMemberships_channelArn :: Lens' BatchChannelMemberships (Maybe Text)

The ARN of the channel to which you're adding users.

batchChannelMemberships_type :: Lens' BatchChannelMemberships (Maybe ChannelMembershipType)

The membership types set for the channel users.

batchChannelMemberships_invitedBy :: Lens' BatchChannelMemberships (Maybe Identity)

The identifier of the member who invited another member.

BatchCreateChannelMembershipError

batchCreateChannelMembershipError_errorCode :: Lens' BatchCreateChannelMembershipError (Maybe ErrorCode)

The error code.

batchCreateChannelMembershipError_memberArn :: Lens' BatchCreateChannelMembershipError (Maybe Text)

The ARN of the member that the service couldn't add.

batchCreateChannelMembershipError_errorMessage :: Lens' BatchCreateChannelMembershipError (Maybe Text)

The error message.

Channel

channel_mode :: Lens' Channel (Maybe ChannelMode)

The mode of the channel.

channel_createdBy :: Lens' Channel (Maybe Identity)

The AppInstanceUser who created the channel.

channel_channelArn :: Lens' Channel (Maybe Text)

The ARN of a channel.

channel_privacy :: Lens' Channel (Maybe ChannelPrivacy)

The channel's privacy setting.

channel_channelFlowArn :: Lens' Channel (Maybe Text)

The ARN of the channel flow.

channel_lastMessageTimestamp :: Lens' Channel (Maybe UTCTime)

The time at which a member sent the last message in the channel.

channel_name :: Lens' Channel (Maybe Text)

The name of a channel.

channel_metadata :: Lens' Channel (Maybe Text)

The channel's metadata.

channel_createdTimestamp :: Lens' Channel (Maybe UTCTime)

The time at which the AppInstanceUser created the channel.

channel_lastUpdatedTimestamp :: Lens' Channel (Maybe UTCTime)

The time at which a channel was last updated.

ChannelAssociatedWithFlowSummary

channelAssociatedWithFlowSummary_mode :: Lens' ChannelAssociatedWithFlowSummary (Maybe ChannelMode)

The mode of the channel.

channelAssociatedWithFlowSummary_channelArn :: Lens' ChannelAssociatedWithFlowSummary (Maybe Text)

The ARN of the channel.

channelAssociatedWithFlowSummary_privacy :: Lens' ChannelAssociatedWithFlowSummary (Maybe ChannelPrivacy)

The channel's privacy setting.

channelAssociatedWithFlowSummary_name :: Lens' ChannelAssociatedWithFlowSummary (Maybe Text)

The name of the channel flow.

channelAssociatedWithFlowSummary_metadata :: Lens' ChannelAssociatedWithFlowSummary (Maybe Text)

The channel's metadata.

ChannelBan

channelBan_createdBy :: Lens' ChannelBan (Maybe Identity)

The AppInstanceUser who created the ban.

channelBan_channelArn :: Lens' ChannelBan (Maybe Text)

The ARN of the channel from which a member is being banned.

channelBan_member :: Lens' ChannelBan (Maybe Identity)

The member being banned from the channel.

channelBan_createdTimestamp :: Lens' ChannelBan (Maybe UTCTime)

The time at which the ban was created.

ChannelBanSummary

channelBanSummary_member :: Lens' ChannelBanSummary (Maybe Identity)

The member being banned from a channel.

ChannelFlow

channelFlow_processors :: Lens' ChannelFlow (Maybe (NonEmpty Processor))

Information about the processor Lambda functions.

channelFlow_channelFlowArn :: Lens' ChannelFlow (Maybe Text)

The ARN of the channel flow.

channelFlow_name :: Lens' ChannelFlow (Maybe Text)

The name of the channel flow.

channelFlow_createdTimestamp :: Lens' ChannelFlow (Maybe UTCTime)

The time at which the channel flow was created.

channelFlow_lastUpdatedTimestamp :: Lens' ChannelFlow (Maybe UTCTime)

The time at which a channel flow was updated.

ChannelFlowSummary

channelFlowSummary_processors :: Lens' ChannelFlowSummary (Maybe (NonEmpty Processor))

Information about the processor Lambda functions.

channelFlowSummary_channelFlowArn :: Lens' ChannelFlowSummary (Maybe Text)

The ARN of the channel flow.

channelFlowSummary_name :: Lens' ChannelFlowSummary (Maybe Text)

The name of the channel flow.

ChannelMembership

channelMembership_channelArn :: Lens' ChannelMembership (Maybe Text)

The ARN of the member's channel.

channelMembership_member :: Lens' ChannelMembership (Maybe Identity)

The data of the channel member.

channelMembership_type :: Lens' ChannelMembership (Maybe ChannelMembershipType)

The membership type set for the channel member.

channelMembership_invitedBy :: Lens' ChannelMembership (Maybe Identity)

The identifier of the member who invited another member.

channelMembership_createdTimestamp :: Lens' ChannelMembership (Maybe UTCTime)

The time at which the channel membership was created.

channelMembership_lastUpdatedTimestamp :: Lens' ChannelMembership (Maybe UTCTime)

The time at which a channel membership was last updated.

ChannelMembershipForAppInstanceUserSummary

channelMembershipForAppInstanceUserSummary_appInstanceUserMembershipSummary :: Lens' ChannelMembershipForAppInstanceUserSummary (Maybe AppInstanceUserMembershipSummary)

Returns the channel membership data for an AppInstance.

channelMembershipForAppInstanceUserSummary_channelSummary :: Lens' ChannelMembershipForAppInstanceUserSummary (Maybe ChannelSummary)

Returns the channel data for an AppInstance.

ChannelMembershipSummary

channelMembershipSummary_member :: Lens' ChannelMembershipSummary (Maybe Identity)

A member's summary data.

ChannelMessage

channelMessage_status :: Lens' ChannelMessage (Maybe ChannelMessageStatusStructure)

The status of the channel message.

channelMessage_sender :: Lens' ChannelMessage (Maybe Identity)

The message sender.

channelMessage_channelArn :: Lens' ChannelMessage (Maybe Text)

The ARN of the channel.

channelMessage_content :: Lens' ChannelMessage (Maybe Text)

The message content.

channelMessage_redacted :: Lens' ChannelMessage (Maybe Bool)

Hides the content of a message.

channelMessage_persistence :: Lens' ChannelMessage (Maybe ChannelMessagePersistenceType)

The persistence setting for a channel message.

channelMessage_metadata :: Lens' ChannelMessage (Maybe Text)

The message metadata.

channelMessage_type :: Lens' ChannelMessage (Maybe ChannelMessageType)

The message type.

channelMessage_createdTimestamp :: Lens' ChannelMessage (Maybe UTCTime)

The time at which the message was created.

channelMessage_messageId :: Lens' ChannelMessage (Maybe Text)

The ID of a message.

channelMessage_lastUpdatedTimestamp :: Lens' ChannelMessage (Maybe UTCTime)

The time at which a message was updated.

channelMessage_lastEditedTimestamp :: Lens' ChannelMessage (Maybe UTCTime)

The time at which a message was edited.

ChannelMessageCallback

channelMessageCallback_content :: Lens' ChannelMessageCallback (Maybe Text)

The message content.

channelMessageCallback_metadata :: Lens' ChannelMessageCallback (Maybe Text)

The message metadata.

channelMessageCallback_messageId :: Lens' ChannelMessageCallback Text

The message ID.

ChannelMessageStatusStructure

channelMessageStatusStructure_value :: Lens' ChannelMessageStatusStructure (Maybe ChannelMessageStatus)

The message status value.

channelMessageStatusStructure_detail :: Lens' ChannelMessageStatusStructure (Maybe Text)

Contains more details about the messasge status.

ChannelMessageSummary

channelMessageSummary_status :: Lens' ChannelMessageSummary (Maybe ChannelMessageStatusStructure)

The message status. The status value is SENT for messages sent to a channel without a channel flow. For channels associated with channel flow, the value determines the processing stage.

channelMessageSummary_sender :: Lens' ChannelMessageSummary (Maybe Identity)

The message sender.

channelMessageSummary_content :: Lens' ChannelMessageSummary (Maybe Text)

The content of the message.

channelMessageSummary_redacted :: Lens' ChannelMessageSummary (Maybe Bool)

Indicates whether a message was redacted.

channelMessageSummary_metadata :: Lens' ChannelMessageSummary (Maybe Text)

The metadata of the message.

channelMessageSummary_type :: Lens' ChannelMessageSummary (Maybe ChannelMessageType)

The type of message.

channelMessageSummary_createdTimestamp :: Lens' ChannelMessageSummary (Maybe UTCTime)

The time at which the message summary was created.

channelMessageSummary_messageId :: Lens' ChannelMessageSummary (Maybe Text)

The ID of the message.

channelMessageSummary_lastUpdatedTimestamp :: Lens' ChannelMessageSummary (Maybe UTCTime)

The time at which a message was last updated.

channelMessageSummary_lastEditedTimestamp :: Lens' ChannelMessageSummary (Maybe UTCTime)

The time at which a message was last edited.

ChannelModeratedByAppInstanceUserSummary

channelModeratedByAppInstanceUserSummary_channelSummary :: Lens' ChannelModeratedByAppInstanceUserSummary (Maybe ChannelSummary)

Summary of the details of a Channel.

ChannelModerator

channelModerator_createdBy :: Lens' ChannelModerator (Maybe Identity)

The AppInstanceUser who created the moderator.

channelModerator_channelArn :: Lens' ChannelModerator (Maybe Text)

The ARN of the moderator's channel.

channelModerator_createdTimestamp :: Lens' ChannelModerator (Maybe UTCTime)

The time at which the moderator was created.

channelModerator_moderator :: Lens' ChannelModerator (Maybe Identity)

The moderator's data.

ChannelModeratorSummary

channelModeratorSummary_moderator :: Lens' ChannelModeratorSummary (Maybe Identity)

The data for a moderator.

ChannelSummary

channelSummary_mode :: Lens' ChannelSummary (Maybe ChannelMode)

The mode of the channel.

channelSummary_channelArn :: Lens' ChannelSummary (Maybe Text)

The ARN of the channel.

channelSummary_privacy :: Lens' ChannelSummary (Maybe ChannelPrivacy)

The privacy setting of the channel.

channelSummary_lastMessageTimestamp :: Lens' ChannelSummary (Maybe UTCTime)

The time at which the last message in a channel was sent.

channelSummary_name :: Lens' ChannelSummary (Maybe Text)

The name of the channel.

channelSummary_metadata :: Lens' ChannelSummary (Maybe Text)

The metadata of the channel.

Identity

identity_arn :: Lens' Identity (Maybe Text)

The ARN in an Identity.

identity_name :: Lens' Identity (Maybe Text)

The name in an Identity.

LambdaConfiguration

lambdaConfiguration_resourceArn :: Lens' LambdaConfiguration Text

The ARN of the Lambda message processing function.

lambdaConfiguration_invocationType :: Lens' LambdaConfiguration InvocationType

Controls how the Lambda function is invoked.

MessagingSessionEndpoint

messagingSessionEndpoint_url :: Lens' MessagingSessionEndpoint (Maybe Text)

The endpoint to which you establish a websocket connection.

Processor

processor_name :: Lens' Processor Text

The name of the channel flow.

processor_configuration :: Lens' Processor ProcessorConfiguration

The information about the type of processor and its identifier.

processor_executionOrder :: Lens' Processor Natural

The sequence in which processors run. If you have multiple processors in a channel flow, message processing goes through each processor in the sequence. The value determines the sequence. At this point, we support only 1 processor within a flow.

processor_fallbackAction :: Lens' Processor FallbackAction

Determines whether to continue or stop processing if communication with processor fails. If the last processor in a channel flow sequence has a fallback action of CONTINUE, and communication with the processor fails, the message is considered processed and sent to the recipients in the channel.

ProcessorConfiguration

processorConfiguration_lambda :: Lens' ProcessorConfiguration LambdaConfiguration

Indicates that the processor is of type Lambda.

Tag

tag_key :: Lens' Tag Text

The key in a tag.

tag_value :: Lens' Tag Text

The value in a tag.