Amazonka.ChimeSDKMessaging.ListChannelFlows

Description

Returns a paginated lists of all the channel flows created under a single Chime. This is a developer API.

Creating a Request

data ListChannelFlows

See: newListChannelFlows smart constructor.

Constructors

ListChannelFlows'

Fields

nextToken :: Maybe (Sensitive Text)

The token passed by previous API calls until all requested channel flows are returned.

maxResults :: Maybe Natural

The maximum number of channel flows that you want to return.

appInstanceArn :: Text

The ARN of the app instance.

newListChannelFlows

Arguments

:: Text
$sel:appInstanceArn:ListChannelFlows'

-> ListChannelFlows

Create a value of ListChannelFlows with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:nextToken:ListChannelFlows', listChannelFlows_nextToken - The token passed by previous API calls until all requested channel flows are returned.

$sel:maxResults:ListChannelFlows', listChannelFlows_maxResults - The maximum number of channel flows that you want to return.

$sel:appInstanceArn:ListChannelFlows', listChannelFlows_appInstanceArn - The ARN of the app instance.

Request Lenses

listChannelFlows_nextToken :: Lens' ListChannelFlows (Maybe Text)

The token passed by previous API calls until all requested channel flows are returned.

listChannelFlows_maxResults :: Lens' ListChannelFlows (Maybe Natural)

The maximum number of channel flows that you want to return.

listChannelFlows_appInstanceArn :: Lens' ListChannelFlows Text

The ARN of the app instance.

Destructuring the Response

data ListChannelFlowsResponse

See: newListChannelFlowsResponse smart constructor.

Constructors

ListChannelFlowsResponse'

Fields

channelFlows :: Maybe [ChannelFlowSummary]

The information about each channel flow.

nextToken :: Maybe (Sensitive Text)

The token passed by previous API calls until all requested channels are returned.

httpStatus :: Int

The response's http status code.

newListChannelFlowsResponse

Arguments

:: Int
$sel:httpStatus:ListChannelFlowsResponse'

-> ListChannelFlowsResponse

Create a value of ListChannelFlowsResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:channelFlows:ListChannelFlowsResponse', listChannelFlowsResponse_channelFlows - The information about each channel flow.

$sel:nextToken:ListChannelFlows', listChannelFlowsResponse_nextToken - The token passed by previous API calls until all requested channels are returned.

$sel:httpStatus:ListChannelFlowsResponse', listChannelFlowsResponse_httpStatus - The response's http status code.

Response Lenses

listChannelFlowsResponse_channelFlows :: Lens' ListChannelFlowsResponse (Maybe [ChannelFlowSummary])

The information about each channel flow.

listChannelFlowsResponse_nextToken :: Lens' ListChannelFlowsResponse (Maybe Text)

The token passed by previous API calls until all requested channels are returned.

listChannelFlowsResponse_httpStatus :: Lens' ListChannelFlowsResponse Int

The response's http status code.