Amazonka.ChimeSDKMessaging.ListChannelMessages

Description

List all the messages in a channel. Returns a paginated list of ChannelMessages. By default, sorted by creation timestamp in descending order.

Redacted messages appear in the results as empty, since they are only redacted, not deleted. Deleted messages do not appear in the results. This action always returns the latest version of an edited message.

Also, the x-amz-chime-bearer request header is mandatory. Use the AppInstanceUserArn of the user that makes the API call as the value in the header.

Creating a Request

data ListChannelMessages

See: newListChannelMessages smart constructor.

Constructors

ListChannelMessages'

Fields

nextToken :: Maybe (Sensitive Text)

The token passed by previous API calls until all requested messages are returned.

notBefore :: Maybe POSIX

The initial or starting time stamp for your requested messages.

sortOrder :: Maybe SortOrder

The order in which you want messages sorted. Default is Descending, based on time created.

maxResults :: Maybe Natural

The maximum number of messages that you want returned.

notAfter :: Maybe POSIX

The final or ending time stamp for your requested messages.

channelArn :: Text

The ARN of the channel.

chimeBearer :: Text

The AppInstanceUserArn of the user that makes the API call.

newListChannelMessages

Arguments

:: Text
$sel:channelArn:ListChannelMessages'

-> Text
$sel:chimeBearer:ListChannelMessages'

-> ListChannelMessages

Create a value of ListChannelMessages with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:nextToken:ListChannelMessages', listChannelMessages_nextToken - The token passed by previous API calls until all requested messages are returned.

$sel:notBefore:ListChannelMessages', listChannelMessages_notBefore - The initial or starting time stamp for your requested messages.

$sel:sortOrder:ListChannelMessages', listChannelMessages_sortOrder - The order in which you want messages sorted. Default is Descending, based on time created.

$sel:maxResults:ListChannelMessages', listChannelMessages_maxResults - The maximum number of messages that you want returned.

$sel:notAfter:ListChannelMessages', listChannelMessages_notAfter - The final or ending time stamp for your requested messages.

$sel:channelArn:ListChannelMessages', listChannelMessages_channelArn - The ARN of the channel.

$sel:chimeBearer:ListChannelMessages', listChannelMessages_chimeBearer - The AppInstanceUserArn of the user that makes the API call.

Request Lenses

listChannelMessages_nextToken :: Lens' ListChannelMessages (Maybe Text)

The token passed by previous API calls until all requested messages are returned.

listChannelMessages_notBefore :: Lens' ListChannelMessages (Maybe UTCTime)

The initial or starting time stamp for your requested messages.

listChannelMessages_sortOrder :: Lens' ListChannelMessages (Maybe SortOrder)

The order in which you want messages sorted. Default is Descending, based on time created.

listChannelMessages_maxResults :: Lens' ListChannelMessages (Maybe Natural)

The maximum number of messages that you want returned.

listChannelMessages_notAfter :: Lens' ListChannelMessages (Maybe UTCTime)

The final or ending time stamp for your requested messages.

listChannelMessages_channelArn :: Lens' ListChannelMessages Text

The ARN of the channel.

listChannelMessages_chimeBearer :: Lens' ListChannelMessages Text

The AppInstanceUserArn of the user that makes the API call.

Destructuring the Response

data ListChannelMessagesResponse

See: newListChannelMessagesResponse smart constructor.

Constructors

ListChannelMessagesResponse'

Fields

channelArn :: Maybe Text

The ARN of the channel containing the requested messages.

nextToken :: Maybe (Sensitive Text)

The token passed by previous API calls until all requested messages are returned.

channelMessages :: Maybe [ChannelMessageSummary]

The information about, and content of, each requested message.

httpStatus :: Int

The response's http status code.

newListChannelMessagesResponse

Arguments

:: Int
$sel:httpStatus:ListChannelMessagesResponse'

-> ListChannelMessagesResponse

Create a value of ListChannelMessagesResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:channelArn:ListChannelMessages', listChannelMessagesResponse_channelArn - The ARN of the channel containing the requested messages.

$sel:nextToken:ListChannelMessages', listChannelMessagesResponse_nextToken - The token passed by previous API calls until all requested messages are returned.

$sel:channelMessages:ListChannelMessagesResponse', listChannelMessagesResponse_channelMessages - The information about, and content of, each requested message.

$sel:httpStatus:ListChannelMessagesResponse', listChannelMessagesResponse_httpStatus - The response's http status code.

Response Lenses

listChannelMessagesResponse_channelArn :: Lens' ListChannelMessagesResponse (Maybe Text)

The ARN of the channel containing the requested messages.

listChannelMessagesResponse_nextToken :: Lens' ListChannelMessagesResponse (Maybe Text)

The token passed by previous API calls until all requested messages are returned.

listChannelMessagesResponse_channelMessages :: Lens' ListChannelMessagesResponse (Maybe [ChannelMessageSummary])

The information about, and content of, each requested message.

listChannelMessagesResponse_httpStatus :: Lens' ListChannelMessagesResponse Int

The response's http status code.