Amazonka.ChimeSDKMessaging.ListChannelModerators

Description

Lists all the moderators for a channel.

The x-amz-chime-bearer request header is mandatory. Use the AppInstanceUserArn of the user that makes the API call as the value in the header.

Creating a Request

data ListChannelModerators

See: newListChannelModerators smart constructor.

Constructors

ListChannelModerators'

Fields

nextToken :: Maybe (Sensitive Text)

The token passed by previous API calls until all requested moderators are returned.

maxResults :: Maybe Natural

The maximum number of moderators that you want returned.

channelArn :: Text

The ARN of the channel.

chimeBearer :: Text

The AppInstanceUserArn of the user that makes the API call.

newListChannelModerators

Arguments

:: Text
$sel:channelArn:ListChannelModerators'

-> Text
$sel:chimeBearer:ListChannelModerators'

-> ListChannelModerators

Create a value of ListChannelModerators with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:nextToken:ListChannelModerators', listChannelModerators_nextToken - The token passed by previous API calls until all requested moderators are returned.

$sel:maxResults:ListChannelModerators', listChannelModerators_maxResults - The maximum number of moderators that you want returned.

$sel:channelArn:ListChannelModerators', listChannelModerators_channelArn - The ARN of the channel.

$sel:chimeBearer:ListChannelModerators', listChannelModerators_chimeBearer - The AppInstanceUserArn of the user that makes the API call.

Request Lenses

listChannelModerators_nextToken :: Lens' ListChannelModerators (Maybe Text)

The token passed by previous API calls until all requested moderators are returned.

listChannelModerators_maxResults :: Lens' ListChannelModerators (Maybe Natural)

The maximum number of moderators that you want returned.

listChannelModerators_channelArn :: Lens' ListChannelModerators Text

The ARN of the channel.

listChannelModerators_chimeBearer :: Lens' ListChannelModerators Text

The AppInstanceUserArn of the user that makes the API call.

Destructuring the Response

data ListChannelModeratorsResponse

See: newListChannelModeratorsResponse smart constructor.

Constructors

ListChannelModeratorsResponse'

Fields

channelArn :: Maybe Text

The ARN of the channel.

nextToken :: Maybe (Sensitive Text)

The token passed by previous API calls until all requested moderators are returned.

channelModerators :: Maybe [ChannelModeratorSummary]

The information about and names of each moderator.

httpStatus :: Int

The response's http status code.

newListChannelModeratorsResponse

Arguments

:: Int
$sel:httpStatus:ListChannelModeratorsResponse'

-> ListChannelModeratorsResponse

Create a value of ListChannelModeratorsResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:channelArn:ListChannelModerators', listChannelModeratorsResponse_channelArn - The ARN of the channel.

$sel:nextToken:ListChannelModerators', listChannelModeratorsResponse_nextToken - The token passed by previous API calls until all requested moderators are returned.

$sel:channelModerators:ListChannelModeratorsResponse', listChannelModeratorsResponse_channelModerators - The information about and names of each moderator.

$sel:httpStatus:ListChannelModeratorsResponse', listChannelModeratorsResponse_httpStatus - The response's http status code.

Response Lenses

listChannelModeratorsResponse_channelArn :: Lens' ListChannelModeratorsResponse (Maybe Text)

The ARN of the channel.

listChannelModeratorsResponse_nextToken :: Lens' ListChannelModeratorsResponse (Maybe Text)

The token passed by previous API calls until all requested moderators are returned.

listChannelModeratorsResponse_channelModerators :: Lens' ListChannelModeratorsResponse (Maybe [ChannelModeratorSummary])

The information about and names of each moderator.

listChannelModeratorsResponse_httpStatus :: Lens' ListChannelModeratorsResponse Int

The response's http status code.