Amazonka.Cloud9.CreateEnvironmentEC2

Description

Creates an Cloud9 development environment, launches an Amazon Elastic Compute Cloud (Amazon EC2) instance, and then connects from the instance to the environment.

Creating a Request

data CreateEnvironmentEC2

See: newCreateEnvironmentEC2 smart constructor.

Constructors

CreateEnvironmentEC2'

Fields

automaticStopTimeMinutes :: Maybe Int

The number of minutes until the running instance is shut down after the environment has last been used.

subnetId :: Maybe Text

The ID of the subnet in Amazon VPC that Cloud9 will use to communicate with the Amazon EC2 instance.

ownerArn :: Maybe Text

The Amazon Resource Name (ARN) of the environment owner. This ARN can be the ARN of any IAM principal. If this value is not specified, the ARN defaults to this environment's creator.

imageId :: Maybe Text

The identifier for the Amazon Machine Image (AMI) that's used to create the EC2 instance. To choose an AMI for the instance, you must specify a valid AMI alias or a valid Amazon EC2 Systems Manager (SSM) path.

The default AMI is used if the parameter isn't explicitly assigned a value in the request. Because Amazon Linux AMI has ended standard support as of December 31, 2020, we recommend you choose Amazon Linux 2, which includes long term support through 2023.

AMI aliases

Amazon Linux (default): amazonlinux-1-x86_64

Amazon Linux 2: amazonlinux-2-x86_64

Ubuntu 18.04: ubuntu-18.04-x86_64

SSM paths

__Amazon Linux (default): resolve:ssm:/aws/service/cloud9/amis/amazonlinux-1-x86_64__

Amazon Linux 2: resolve:ssm:/aws/service/cloud9/amis/amazonlinux-2-x86_64

Ubuntu 18.04: resolve:ssm:/aws/service/cloud9/amis/ubuntu-18.04-x86_64

clientRequestToken :: Maybe Text

A unique, case-sensitive string that helps Cloud9 to ensure this operation completes no more than one time.

For more information, see Client Tokens in the Amazon EC2 API Reference.

connectionType :: Maybe ConnectionType

The connection type used for connecting to an Amazon EC2 environment. Valid values are CONNECT_SSH (default) and CONNECT_SSM (connected through Amazon EC2 Systems Manager).

For more information, see Accessing no-ingress EC2 instances with Amazon EC2 Systems Manager in the Cloud9 User Guide.

description :: Maybe (Sensitive Text)

The description of the environment to create.

dryRun :: Maybe Bool

Checks whether you have the required permissions for the action, without actually making the request, and provides an error response. If you have the required permissions, the error response is DryRunOperation. Otherwise, it is UnauthorizedOperation.

tags :: Maybe (Sensitive [Sensitive Tag])

An array of key-value pairs that will be associated with the new Cloud9 development environment.

name :: Text

The name of the environment to create.

This name is visible to other IAM users in the same Amazon Web Services account.

instanceType :: Text

The type of instance to connect to the environment (for example, t2.micro).

newCreateEnvironmentEC2

Arguments

:: Text
$sel:name:CreateEnvironmentEC2'

-> Text
$sel:instanceType:CreateEnvironmentEC2'

-> CreateEnvironmentEC2

Create a value of CreateEnvironmentEC2 with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:automaticStopTimeMinutes:CreateEnvironmentEC2', createEnvironmentEC2_automaticStopTimeMinutes - The number of minutes until the running instance is shut down after the environment has last been used.

$sel:subnetId:CreateEnvironmentEC2', createEnvironmentEC2_subnetId - The ID of the subnet in Amazon VPC that Cloud9 will use to communicate with the Amazon EC2 instance.

$sel:ownerArn:CreateEnvironmentEC2', createEnvironmentEC2_ownerArn - The Amazon Resource Name (ARN) of the environment owner. This ARN can be the ARN of any IAM principal. If this value is not specified, the ARN defaults to this environment's creator.

$sel:imageId:CreateEnvironmentEC2', createEnvironmentEC2_imageId - The identifier for the Amazon Machine Image (AMI) that's used to create the EC2 instance. To choose an AMI for the instance, you must specify a valid AMI alias or a valid Amazon EC2 Systems Manager (SSM) path.

The default AMI is used if the parameter isn't explicitly assigned a value in the request. Because Amazon Linux AMI has ended standard support as of December 31, 2020, we recommend you choose Amazon Linux 2, which includes long term support through 2023.

AMI aliases

Amazon Linux (default): amazonlinux-1-x86_64

Amazon Linux 2: amazonlinux-2-x86_64

Ubuntu 18.04: ubuntu-18.04-x86_64

SSM paths

__Amazon Linux (default): resolve:ssm:/aws/service/cloud9/amis/amazonlinux-1-x86_64__

Amazon Linux 2: resolve:ssm:/aws/service/cloud9/amis/amazonlinux-2-x86_64

Ubuntu 18.04: resolve:ssm:/aws/service/cloud9/amis/ubuntu-18.04-x86_64

$sel:clientRequestToken:CreateEnvironmentEC2', createEnvironmentEC2_clientRequestToken - A unique, case-sensitive string that helps Cloud9 to ensure this operation completes no more than one time.

For more information, see Client Tokens in the Amazon EC2 API Reference.

$sel:connectionType:CreateEnvironmentEC2', createEnvironmentEC2_connectionType - The connection type used for connecting to an Amazon EC2 environment. Valid values are CONNECT_SSH (default) and CONNECT_SSM (connected through Amazon EC2 Systems Manager).

For more information, see Accessing no-ingress EC2 instances with Amazon EC2 Systems Manager in the Cloud9 User Guide.

$sel:description:CreateEnvironmentEC2', createEnvironmentEC2_description - The description of the environment to create.

$sel:dryRun:CreateEnvironmentEC2', createEnvironmentEC2_dryRun - Checks whether you have the required permissions for the action, without actually making the request, and provides an error response. If you have the required permissions, the error response is DryRunOperation. Otherwise, it is UnauthorizedOperation.

$sel:tags:CreateEnvironmentEC2', createEnvironmentEC2_tags - An array of key-value pairs that will be associated with the new Cloud9 development environment.

$sel:name:CreateEnvironmentEC2', createEnvironmentEC2_name - The name of the environment to create.

This name is visible to other IAM users in the same Amazon Web Services account.

$sel:instanceType:CreateEnvironmentEC2', createEnvironmentEC2_instanceType - The type of instance to connect to the environment (for example, t2.micro).

Request Lenses

createEnvironmentEC2_automaticStopTimeMinutes :: Lens' CreateEnvironmentEC2 (Maybe Int)

The number of minutes until the running instance is shut down after the environment has last been used.

createEnvironmentEC2_subnetId :: Lens' CreateEnvironmentEC2 (Maybe Text)

The ID of the subnet in Amazon VPC that Cloud9 will use to communicate with the Amazon EC2 instance.

createEnvironmentEC2_ownerArn :: Lens' CreateEnvironmentEC2 (Maybe Text)

The Amazon Resource Name (ARN) of the environment owner. This ARN can be the ARN of any IAM principal. If this value is not specified, the ARN defaults to this environment's creator.

createEnvironmentEC2_imageId :: Lens' CreateEnvironmentEC2 (Maybe Text)

The identifier for the Amazon Machine Image (AMI) that's used to create the EC2 instance. To choose an AMI for the instance, you must specify a valid AMI alias or a valid Amazon EC2 Systems Manager (SSM) path.

The default AMI is used if the parameter isn't explicitly assigned a value in the request. Because Amazon Linux AMI has ended standard support as of December 31, 2020, we recommend you choose Amazon Linux 2, which includes long term support through 2023.

AMI aliases

Amazon Linux (default): amazonlinux-1-x86_64

Amazon Linux 2: amazonlinux-2-x86_64

Ubuntu 18.04: ubuntu-18.04-x86_64

SSM paths

__Amazon Linux (default): resolve:ssm:/aws/service/cloud9/amis/amazonlinux-1-x86_64__

Amazon Linux 2: resolve:ssm:/aws/service/cloud9/amis/amazonlinux-2-x86_64

Ubuntu 18.04: resolve:ssm:/aws/service/cloud9/amis/ubuntu-18.04-x86_64

createEnvironmentEC2_clientRequestToken :: Lens' CreateEnvironmentEC2 (Maybe Text)

A unique, case-sensitive string that helps Cloud9 to ensure this operation completes no more than one time.

For more information, see Client Tokens in the Amazon EC2 API Reference.

createEnvironmentEC2_connectionType :: Lens' CreateEnvironmentEC2 (Maybe ConnectionType)

The connection type used for connecting to an Amazon EC2 environment. Valid values are CONNECT_SSH (default) and CONNECT_SSM (connected through Amazon EC2 Systems Manager).

For more information, see Accessing no-ingress EC2 instances with Amazon EC2 Systems Manager in the Cloud9 User Guide.

createEnvironmentEC2_description :: Lens' CreateEnvironmentEC2 (Maybe Text)

The description of the environment to create.

createEnvironmentEC2_dryRun :: Lens' CreateEnvironmentEC2 (Maybe Bool)

Checks whether you have the required permissions for the action, without actually making the request, and provides an error response. If you have the required permissions, the error response is DryRunOperation. Otherwise, it is UnauthorizedOperation.

createEnvironmentEC2_tags :: Lens' CreateEnvironmentEC2 (Maybe [Tag])

An array of key-value pairs that will be associated with the new Cloud9 development environment.

createEnvironmentEC2_name :: Lens' CreateEnvironmentEC2 Text

The name of the environment to create.

This name is visible to other IAM users in the same Amazon Web Services account.

createEnvironmentEC2_instanceType :: Lens' CreateEnvironmentEC2 Text

The type of instance to connect to the environment (for example, t2.micro).

Destructuring the Response

data CreateEnvironmentEC2Response

See: newCreateEnvironmentEC2Response smart constructor.

Constructors

CreateEnvironmentEC2Response'

Fields

environmentId :: Maybe Text

The ID of the environment that was created.

httpStatus :: Int

The response's http status code.

newCreateEnvironmentEC2Response

Arguments

:: Int
$sel:httpStatus:CreateEnvironmentEC2Response'

-> CreateEnvironmentEC2Response

Create a value of CreateEnvironmentEC2Response with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:environmentId:CreateEnvironmentEC2Response', createEnvironmentEC2Response_environmentId - The ID of the environment that was created.

$sel:httpStatus:CreateEnvironmentEC2Response', createEnvironmentEC2Response_httpStatus - The response's http status code.

Response Lenses

createEnvironmentEC2Response_environmentId :: Lens' CreateEnvironmentEC2Response (Maybe Text)

The ID of the environment that was created.

createEnvironmentEC2Response_httpStatus :: Lens' CreateEnvironmentEC2Response Int

The response's http status code.