Amazonka.Cloud9.ListEnvironments

Description

Gets a list of Cloud9 development environment identifiers.

This operation returns paginated results.

Creating a Request

data ListEnvironments

See: newListEnvironments smart constructor.

Constructors

ListEnvironments'

Fields

nextToken :: Maybe Text

During a previous call, if there are more than 25 items in the list, only the first 25 items are returned, along with a unique string called a next token. To get the next batch of items in the list, call this operation again, adding the next token to the call. To get all of the items in the list, keep calling this operation with each subsequent next token that is returned, until no more next tokens are returned.

maxResults :: Maybe Natural

The maximum number of environments to get identifiers for.

newListEnvironments :: ListEnvironments

Create a value of ListEnvironments with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:nextToken:ListEnvironments', listEnvironments_nextToken - During a previous call, if there are more than 25 items in the list, only the first 25 items are returned, along with a unique string called a next token. To get the next batch of items in the list, call this operation again, adding the next token to the call. To get all of the items in the list, keep calling this operation with each subsequent next token that is returned, until no more next tokens are returned.

$sel:maxResults:ListEnvironments', listEnvironments_maxResults - The maximum number of environments to get identifiers for.

Request Lenses

listEnvironments_nextToken :: Lens' ListEnvironments (Maybe Text)

During a previous call, if there are more than 25 items in the list, only the first 25 items are returned, along with a unique string called a next token. To get the next batch of items in the list, call this operation again, adding the next token to the call. To get all of the items in the list, keep calling this operation with each subsequent next token that is returned, until no more next tokens are returned.

listEnvironments_maxResults :: Lens' ListEnvironments (Maybe Natural)

The maximum number of environments to get identifiers for.

Destructuring the Response

data ListEnvironmentsResponse

See: newListEnvironmentsResponse smart constructor.

Constructors

ListEnvironmentsResponse'

Fields

environmentIds :: Maybe [Text]

The list of environment identifiers.

nextToken :: Maybe Text

If there are more than 25 items in the list, only the first 25 items are returned, along with a unique string called a next token. To get the next batch of items in the list, call this operation again, adding the next token to the call.

httpStatus :: Int

The response's http status code.

newListEnvironmentsResponse

Arguments

:: Int
$sel:httpStatus:ListEnvironmentsResponse'

-> ListEnvironmentsResponse

Create a value of ListEnvironmentsResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:environmentIds:ListEnvironmentsResponse', listEnvironmentsResponse_environmentIds - The list of environment identifiers.

$sel:nextToken:ListEnvironments', listEnvironmentsResponse_nextToken - If there are more than 25 items in the list, only the first 25 items are returned, along with a unique string called a next token. To get the next batch of items in the list, call this operation again, adding the next token to the call.

$sel:httpStatus:ListEnvironmentsResponse', listEnvironmentsResponse_httpStatus - The response's http status code.

Response Lenses

listEnvironmentsResponse_environmentIds :: Lens' ListEnvironmentsResponse (Maybe [Text])

The list of environment identifiers.

listEnvironmentsResponse_nextToken :: Lens' ListEnvironmentsResponse (Maybe Text)

If there are more than 25 items in the list, only the first 25 items are returned, along with a unique string called a next token. To get the next batch of items in the list, call this operation again, adding the next token to the call.

listEnvironmentsResponse_httpStatus :: Lens' ListEnvironmentsResponse Int

The response's http status code.