Amazonka.Cloud9.Types

Description

Service Configuration

defaultService :: Service

API version 2017-09-23 of the Amazon Cloud9 SDK configuration.

Errors

_ConflictException :: AsError a => Getting (First ServiceError) a ServiceError

A conflict occurred.

_ForbiddenException :: AsError a => Getting (First ServiceError) a ServiceError

An access permissions issue occurred.

_NotFoundException :: AsError a => Getting (First ServiceError) a ServiceError

The target resource cannot be found.

_TooManyRequestsException :: AsError a => Getting (First ServiceError) a ServiceError

Too many service requests were made over the given time period.

_InternalServerErrorException :: AsError a => Getting (First ServiceError) a ServiceError

An internal server error occurred.

_ConcurrentAccessException :: AsError a => Getting (First ServiceError) a ServiceError

A concurrent access issue occurred.

_BadRequestException :: AsError a => Getting (First ServiceError) a ServiceError

The target request is invalid.

_LimitExceededException :: AsError a => Getting (First ServiceError) a ServiceError

A service limit was exceeded.

ConnectionType

newtype ConnectionType

Constructors

ConnectionType'

Fields

fromConnectionType :: Text

Bundled Patterns

pattern ConnectionType_CONNECT_SSH :: ConnectionType

pattern ConnectionType_CONNECT_SSM :: ConnectionType

EnvironmentLifecycleStatus

newtype EnvironmentLifecycleStatus

Constructors

EnvironmentLifecycleStatus'

Fields

fromEnvironmentLifecycleStatus :: Text

Bundled Patterns

pattern EnvironmentLifecycleStatus_CREATED :: EnvironmentLifecycleStatus

pattern EnvironmentLifecycleStatus_CREATE_FAILED :: EnvironmentLifecycleStatus

pattern EnvironmentLifecycleStatus_CREATING :: EnvironmentLifecycleStatus

pattern EnvironmentLifecycleStatus_DELETE_FAILED :: EnvironmentLifecycleStatus

pattern EnvironmentLifecycleStatus_DELETING :: EnvironmentLifecycleStatus

EnvironmentStatus

newtype EnvironmentStatus

Constructors

EnvironmentStatus'

Fields

fromEnvironmentStatus :: Text

Bundled Patterns

pattern EnvironmentStatus_Connecting :: EnvironmentStatus

pattern EnvironmentStatus_Creating :: EnvironmentStatus

pattern EnvironmentStatus_Deleting :: EnvironmentStatus

pattern EnvironmentStatus_Error :: EnvironmentStatus

pattern EnvironmentStatus_Ready :: EnvironmentStatus

pattern EnvironmentStatus_Stopped :: EnvironmentStatus

pattern EnvironmentStatus_Stopping :: EnvironmentStatus

EnvironmentType

newtype EnvironmentType

Constructors

EnvironmentType'

Fields

fromEnvironmentType :: Text

Bundled Patterns

pattern EnvironmentType_Ec2 :: EnvironmentType

pattern EnvironmentType_Ssh :: EnvironmentType

ManagedCredentialsAction

newtype ManagedCredentialsAction

Constructors

ManagedCredentialsAction'

Fields

fromManagedCredentialsAction :: Text

Bundled Patterns

pattern ManagedCredentialsAction_DISABLE :: ManagedCredentialsAction

pattern ManagedCredentialsAction_ENABLE :: ManagedCredentialsAction

ManagedCredentialsStatus

newtype ManagedCredentialsStatus

Constructors

ManagedCredentialsStatus'

Fields

fromManagedCredentialsStatus :: Text

Bundled Patterns

pattern ManagedCredentialsStatus_DISABLED_BY_COLLABORATOR :: ManagedCredentialsStatus

pattern ManagedCredentialsStatus_DISABLED_BY_DEFAULT :: ManagedCredentialsStatus

pattern ManagedCredentialsStatus_DISABLED_BY_OWNER :: ManagedCredentialsStatus

pattern ManagedCredentialsStatus_ENABLED_BY_OWNER :: ManagedCredentialsStatus

pattern ManagedCredentialsStatus_ENABLED_ON_CREATE :: ManagedCredentialsStatus

pattern ManagedCredentialsStatus_FAILED_REMOVAL_BY_COLLABORATOR :: ManagedCredentialsStatus

pattern ManagedCredentialsStatus_FAILED_REMOVAL_BY_OWNER :: ManagedCredentialsStatus

pattern ManagedCredentialsStatus_PENDING_REMOVAL_BY_COLLABORATOR :: ManagedCredentialsStatus

pattern ManagedCredentialsStatus_PENDING_REMOVAL_BY_OWNER :: ManagedCredentialsStatus

pattern ManagedCredentialsStatus_PENDING_START_REMOVAL_BY_COLLABORATOR :: ManagedCredentialsStatus

pattern ManagedCredentialsStatus_PENDING_START_REMOVAL_BY_OWNER :: ManagedCredentialsStatus

MemberPermissions

newtype MemberPermissions

Constructors

MemberPermissions'

Fields

fromMemberPermissions :: Text

Bundled Patterns

pattern MemberPermissions_Read_only :: MemberPermissions

pattern MemberPermissions_Read_write :: MemberPermissions

Permissions

newtype Permissions

Constructors

Permissions'

Fields

fromPermissions :: Text

Bundled Patterns

pattern Permissions_Owner :: Permissions

pattern Permissions_Read_only :: Permissions

pattern Permissions_Read_write :: Permissions

Environment

data Environment

Information about an Cloud9 development environment.

See: newEnvironment smart constructor.

Constructors

Environment'

Fields

lifecycle :: Maybe EnvironmentLifecycle

The state of the environment in its creation or deletion lifecycle.

name :: Maybe Text

The name of the environment.

id :: Maybe Text

The ID of the environment.

connectionType :: Maybe ConnectionType

The connection type used for connecting to an Amazon EC2 environment. CONNECT_SSH is selected by default.

description :: Maybe (Sensitive Text)

The description for the environment.

managedCredentialsStatus :: Maybe ManagedCredentialsStatus

Describes the status of Amazon Web Services managed temporary credentials for the Cloud9 environment. Available values are:

ENABLED_ON_CREATE

ENABLED_BY_OWNER

DISABLED_BY_DEFAULT

DISABLED_BY_OWNER

DISABLED_BY_COLLABORATOR

PENDING_REMOVAL_BY_COLLABORATOR

PENDING_REMOVAL_BY_OWNER

FAILED_REMOVAL_BY_COLLABORATOR

ENABLED_BY_OWNER

DISABLED_BY_DEFAULT

type' :: EnvironmentType

The type of environment. Valid values include the following:

ec2: An Amazon Elastic Compute Cloud (Amazon EC2) instance connects to the environment.

ssh: Your own server connects to the environment.

arn :: Text

The Amazon Resource Name (ARN) of the environment.

ownerArn :: Text

The Amazon Resource Name (ARN) of the environment owner.

newEnvironment

Arguments

:: EnvironmentType
$sel:type':Environment'

-> Text
$sel:arn:Environment'

-> Text
$sel:ownerArn:Environment'

-> Environment

Create a value of Environment with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:lifecycle:Environment', environment_lifecycle - The state of the environment in its creation or deletion lifecycle.

$sel:name:Environment', environment_name - The name of the environment.

$sel:id:Environment', environment_id - The ID of the environment.

$sel:connectionType:Environment', environment_connectionType - The connection type used for connecting to an Amazon EC2 environment. CONNECT_SSH is selected by default.

$sel:description:Environment', environment_description - The description for the environment.

$sel:managedCredentialsStatus:Environment', environment_managedCredentialsStatus - Describes the status of Amazon Web Services managed temporary credentials for the Cloud9 environment. Available values are:

ENABLED_ON_CREATE

ENABLED_BY_OWNER

DISABLED_BY_DEFAULT

DISABLED_BY_OWNER

DISABLED_BY_COLLABORATOR

PENDING_REMOVAL_BY_COLLABORATOR

PENDING_REMOVAL_BY_OWNER

FAILED_REMOVAL_BY_COLLABORATOR

ENABLED_BY_OWNER

DISABLED_BY_DEFAULT

$sel:type':Environment', environment_type - The type of environment. Valid values include the following:

ec2: An Amazon Elastic Compute Cloud (Amazon EC2) instance connects to the environment.

ssh: Your own server connects to the environment.

$sel:arn:Environment', environment_arn - The Amazon Resource Name (ARN) of the environment.

$sel:ownerArn:Environment', environment_ownerArn - The Amazon Resource Name (ARN) of the environment owner.

environment_lifecycle :: Lens' Environment (Maybe EnvironmentLifecycle)

The state of the environment in its creation or deletion lifecycle.

environment_name :: Lens' Environment (Maybe Text)

The name of the environment.

environment_id :: Lens' Environment (Maybe Text)

The ID of the environment.

environment_connectionType :: Lens' Environment (Maybe ConnectionType)

The connection type used for connecting to an Amazon EC2 environment. CONNECT_SSH is selected by default.

environment_description :: Lens' Environment (Maybe Text)

The description for the environment.

environment_managedCredentialsStatus :: Lens' Environment (Maybe ManagedCredentialsStatus)

Describes the status of Amazon Web Services managed temporary credentials for the Cloud9 environment. Available values are:

ENABLED_ON_CREATE

ENABLED_BY_OWNER

DISABLED_BY_DEFAULT

DISABLED_BY_OWNER

DISABLED_BY_COLLABORATOR

PENDING_REMOVAL_BY_COLLABORATOR

PENDING_REMOVAL_BY_OWNER

FAILED_REMOVAL_BY_COLLABORATOR

ENABLED_BY_OWNER

DISABLED_BY_DEFAULT

environment_type :: Lens' Environment EnvironmentType

The type of environment. Valid values include the following:

ec2: An Amazon Elastic Compute Cloud (Amazon EC2) instance connects to the environment.

ssh: Your own server connects to the environment.

environment_arn :: Lens' Environment Text

The Amazon Resource Name (ARN) of the environment.

environment_ownerArn :: Lens' Environment Text

The Amazon Resource Name (ARN) of the environment owner.

EnvironmentLifecycle

data EnvironmentLifecycle

Information about the current creation or deletion lifecycle state of an Cloud9 development environment.

See: newEnvironmentLifecycle smart constructor.

Constructors

EnvironmentLifecycle'

Fields

status :: Maybe EnvironmentLifecycleStatus

The current creation or deletion lifecycle state of the environment.

CREATING: The environment is in the process of being created.

CREATED: The environment was successfully created.

CREATE_FAILED: The environment failed to be created.

DELETING: The environment is in the process of being deleted.

DELETE_FAILED: The environment failed to delete.

failureResource :: Maybe Text

If the environment failed to delete, the Amazon Resource Name (ARN) of the related Amazon Web Services resource.

reason :: Maybe Text

Any informational message about the lifecycle state of the environment.

newEnvironmentLifecycle :: EnvironmentLifecycle

Create a value of EnvironmentLifecycle with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:status:EnvironmentLifecycle', environmentLifecycle_status - The current creation or deletion lifecycle state of the environment.

CREATING: The environment is in the process of being created.

CREATED: The environment was successfully created.

CREATE_FAILED: The environment failed to be created.

DELETING: The environment is in the process of being deleted.

DELETE_FAILED: The environment failed to delete.

$sel:failureResource:EnvironmentLifecycle', environmentLifecycle_failureResource - If the environment failed to delete, the Amazon Resource Name (ARN) of the related Amazon Web Services resource.

$sel:reason:EnvironmentLifecycle', environmentLifecycle_reason - Any informational message about the lifecycle state of the environment.

environmentLifecycle_status :: Lens' EnvironmentLifecycle (Maybe EnvironmentLifecycleStatus)

The current creation or deletion lifecycle state of the environment.

CREATING: The environment is in the process of being created.

CREATED: The environment was successfully created.

CREATE_FAILED: The environment failed to be created.

DELETING: The environment is in the process of being deleted.

DELETE_FAILED: The environment failed to delete.

environmentLifecycle_failureResource :: Lens' EnvironmentLifecycle (Maybe Text)

If the environment failed to delete, the Amazon Resource Name (ARN) of the related Amazon Web Services resource.

environmentLifecycle_reason :: Lens' EnvironmentLifecycle (Maybe Text)

Any informational message about the lifecycle state of the environment.

EnvironmentMember

data EnvironmentMember

Information about an environment member for an Cloud9 development environment.

See: newEnvironmentMember smart constructor.

Constructors

EnvironmentMember'

Fields

lastAccess :: Maybe POSIX

The time, expressed in epoch time format, when the environment member last opened the environment.

permissions :: Permissions

The type of environment member permissions associated with this environment member. Available values include:

owner: Owns the environment.

read-only: Has read-only access to the environment.

read-write: Has read-write access to the environment.

userId :: Text

The user ID in Identity and Access Management (IAM) of the environment member.

userArn :: Text

The Amazon Resource Name (ARN) of the environment member.

environmentId :: Text

The ID of the environment for the environment member.

newEnvironmentMember

Arguments

:: Permissions
$sel:permissions:EnvironmentMember'

-> Text
$sel:userId:EnvironmentMember'

-> Text
$sel:userArn:EnvironmentMember'

-> Text
$sel:environmentId:EnvironmentMember'

-> EnvironmentMember

Create a value of EnvironmentMember with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:lastAccess:EnvironmentMember', environmentMember_lastAccess - The time, expressed in epoch time format, when the environment member last opened the environment.

$sel:permissions:EnvironmentMember', environmentMember_permissions - The type of environment member permissions associated with this environment member. Available values include:

owner: Owns the environment.

read-only: Has read-only access to the environment.

read-write: Has read-write access to the environment.

$sel:userId:EnvironmentMember', environmentMember_userId - The user ID in Identity and Access Management (IAM) of the environment member.

$sel:userArn:EnvironmentMember', environmentMember_userArn - The Amazon Resource Name (ARN) of the environment member.

$sel:environmentId:EnvironmentMember', environmentMember_environmentId - The ID of the environment for the environment member.

environmentMember_lastAccess :: Lens' EnvironmentMember (Maybe UTCTime)

The time, expressed in epoch time format, when the environment member last opened the environment.

environmentMember_permissions :: Lens' EnvironmentMember Permissions

The type of environment member permissions associated with this environment member. Available values include:

owner: Owns the environment.

read-only: Has read-only access to the environment.

read-write: Has read-write access to the environment.

environmentMember_userId :: Lens' EnvironmentMember Text

The user ID in Identity and Access Management (IAM) of the environment member.

environmentMember_userArn :: Lens' EnvironmentMember Text

The Amazon Resource Name (ARN) of the environment member.

environmentMember_environmentId :: Lens' EnvironmentMember Text

The ID of the environment for the environment member.

Tag

data Tag

Metadata that is associated with Amazon Web Services resources. In particular, a name-value pair that can be associated with an Cloud9 development environment. There are two types of tags: user tags and system tags. A user tag is created by the user. A system tag is automatically created by Amazon Web Services services. A system tag is prefixed with "aws:" and cannot be modified by the user.

See: newTag smart constructor.

Constructors

Tag'

Fields

key :: Sensitive Text

The name part of a tag.

value :: Sensitive Text

The value part of a tag.

newTag

Arguments

:: Text
$sel:key:Tag'

-> Text
$sel:value:Tag'

-> Tag

Create a value of Tag with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:key:Tag', tag_key - The name part of a tag.

$sel:value:Tag', tag_value - The value part of a tag.

tag_key :: Lens' Tag Text

The name part of a tag.

tag_value :: Lens' Tag Text

The value part of a tag.