Amazonka.Cloud9.UpdateEnvironmentMembership

Description

Changes the settings of an existing environment member for an Cloud9 development environment.

Creating a Request

data UpdateEnvironmentMembership

See: newUpdateEnvironmentMembership smart constructor.

Constructors

UpdateEnvironmentMembership'

Fields

environmentId :: Text

The ID of the environment for the environment member whose settings you want to change.

userArn :: Text

The Amazon Resource Name (ARN) of the environment member whose settings you want to change.

permissions :: MemberPermissions

The replacement type of environment member permissions you want to associate with this environment member. Available values include:

read-only: Has read-only access to the environment.

read-write: Has read-write access to the environment.

newUpdateEnvironmentMembership

Arguments

:: Text
$sel:environmentId:UpdateEnvironmentMembership'

-> Text
$sel:userArn:UpdateEnvironmentMembership'

-> MemberPermissions
$sel:permissions:UpdateEnvironmentMembership'

-> UpdateEnvironmentMembership

Create a value of UpdateEnvironmentMembership with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:environmentId:UpdateEnvironmentMembership', updateEnvironmentMembership_environmentId - The ID of the environment for the environment member whose settings you want to change.

$sel:userArn:UpdateEnvironmentMembership', updateEnvironmentMembership_userArn - The Amazon Resource Name (ARN) of the environment member whose settings you want to change.

$sel:permissions:UpdateEnvironmentMembership', updateEnvironmentMembership_permissions - The replacement type of environment member permissions you want to associate with this environment member. Available values include:

read-only: Has read-only access to the environment.

read-write: Has read-write access to the environment.

Request Lenses

updateEnvironmentMembership_environmentId :: Lens' UpdateEnvironmentMembership Text

The ID of the environment for the environment member whose settings you want to change.

updateEnvironmentMembership_userArn :: Lens' UpdateEnvironmentMembership Text

The Amazon Resource Name (ARN) of the environment member whose settings you want to change.

updateEnvironmentMembership_permissions :: Lens' UpdateEnvironmentMembership MemberPermissions

The replacement type of environment member permissions you want to associate with this environment member. Available values include:

read-only: Has read-only access to the environment.

read-write: Has read-write access to the environment.

Destructuring the Response

data UpdateEnvironmentMembershipResponse

See: newUpdateEnvironmentMembershipResponse smart constructor.

Constructors

UpdateEnvironmentMembershipResponse'

Fields

membership :: Maybe EnvironmentMember

Information about the environment member whose settings were changed.

httpStatus :: Int

The response's http status code.

newUpdateEnvironmentMembershipResponse

Arguments

:: Int
$sel:httpStatus:UpdateEnvironmentMembershipResponse'

-> UpdateEnvironmentMembershipResponse

Create a value of UpdateEnvironmentMembershipResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:membership:UpdateEnvironmentMembershipResponse', updateEnvironmentMembershipResponse_membership - Information about the environment member whose settings were changed.

$sel:httpStatus:UpdateEnvironmentMembershipResponse', updateEnvironmentMembershipResponse_httpStatus - The response's http status code.

Response Lenses

updateEnvironmentMembershipResponse_membership :: Lens' UpdateEnvironmentMembershipResponse (Maybe EnvironmentMember)

Information about the environment member whose settings were changed.

updateEnvironmentMembershipResponse_httpStatus :: Lens' UpdateEnvironmentMembershipResponse Int

The response's http status code.