Amazonka.CloudControl

Description

Derived from API version 2021-09-30 of the AWS service descriptions, licensed under Apache 2.0.

Use Amazon Web Services Cloud Control API to create, read, update, delete, and list (CRUD-L) your cloud resources that belong to a wide range of services--both Amazon Web Services and third-party. With the Cloud Control API standardized set of application programming interfaces (APIs), you can perform CRUD-L operations on any supported resources in your Amazon Web Services account. Using Cloud Control API, you won't have to generate code or scripts specific to each individual service responsible for those resources.

For more information about Amazon Web Services Cloud Control API, see the Amazon Web Services Cloud Control API User Guide.

Service Configuration

defaultService :: Service

API version 2021-09-30 of the Amazon Cloud Control API SDK configuration.

Errors

Error matchers are designed for use with the functions provided by Control.Exception.Lens. This allows catching (and rethrowing) service specific errors returned by CloudControl.

TypeNotFoundException

_TypeNotFoundException :: AsError a => Getting (First ServiceError) a ServiceError

The specified extension does not exist in the CloudFormation registry.

GeneralServiceException

_GeneralServiceException :: AsError a => Getting (First ServiceError) a ServiceError

The resource handler has returned that the downstream service generated an error that does not map to any other handler error code.

HandlerFailureException

_HandlerFailureException :: AsError a => Getting (First ServiceError) a ServiceError

The resource handler has failed without a returning a more specific error code. This can include timeouts.

InvalidRequestException

_InvalidRequestException :: AsError a => Getting (First ServiceError) a ServiceError

The resource handler has returned that invalid input from the user has generated a generic exception.

PrivateTypeException

_PrivateTypeException :: AsError a => Getting (First ServiceError) a ServiceError

Cloud Control API has not received a valid response from the resource handler, due to a configuration error. This includes issues such as the resource handler returning an invalid response, or timing out.

ServiceInternalErrorException

_ServiceInternalErrorException :: AsError a => Getting (First ServiceError) a ServiceError

The resource handler has returned that the downstream service returned an internal error, typically with a 5XX HTTP status code.

HandlerInternalFailureException

_HandlerInternalFailureException :: AsError a => Getting (First ServiceError) a ServiceError

The resource handler has returned that an unexpected error occurred within the resource handler.

ClientTokenConflictException

_ClientTokenConflictException :: AsError a => Getting (First ServiceError) a ServiceError

The specified client token has already been used in another resource request.

It is best practice for client tokens to be unique for each resource operation request. However, client token expire after 36 hours.

RequestTokenNotFoundException

_RequestTokenNotFoundException :: AsError a => Getting (First ServiceError) a ServiceError

A resource operation with the specified request token cannot be found.

NetworkFailureException

_NetworkFailureException :: AsError a => Getting (First ServiceError) a ServiceError

The resource handler has returned that the request could not be completed due to networking issues, such as a failure to receive a response from the server.

NotStabilizedException

_NotStabilizedException :: AsError a => Getting (First ServiceError) a ServiceError

The resource handler has returned that the downstream resource failed to complete all of its ready-state checks.

NotUpdatableException

_NotUpdatableException :: AsError a => Getting (First ServiceError) a ServiceError

One or more properties included in this resource operation are defined as create-only, and therefore cannot be updated.

UnsupportedActionException

_UnsupportedActionException :: AsError a => Getting (First ServiceError) a ServiceError

The specified resource does not support this resource operation.

ServiceLimitExceededException

_ServiceLimitExceededException :: AsError a => Getting (First ServiceError) a ServiceError

The resource handler has returned that a non-transient resource limit was reached on the service side.

ThrottlingException

_ThrottlingException :: AsError a => Getting (First ServiceError) a ServiceError

The request was denied due to request throttling.

ConcurrentModificationException

_ConcurrentModificationException :: AsError a => Getting (First ServiceError) a ServiceError

The resource is currently being modified by another operation.

ResourceConflictException

_ResourceConflictException :: AsError a => Getting (First ServiceError) a ServiceError

The resource is temporarily unavailable to be acted upon. For example, if the resource is currently undergoing an operation and cannot be acted upon until that operation is finished.

InvalidCredentialsException

_InvalidCredentialsException :: AsError a => Getting (First ServiceError) a ServiceError

The resource handler has returned that the credentials provided by the user are invalid.

ResourceNotFoundException

_ResourceNotFoundException :: AsError a => Getting (First ServiceError) a ServiceError

A resource with the specified identifier cannot be found.

ConcurrentOperationException

_ConcurrentOperationException :: AsError a => Getting (First ServiceError) a ServiceError

Another resource operation is currently being performed on this resource.

AlreadyExistsException

_AlreadyExistsException :: AsError a => Getting (First ServiceError) a ServiceError

The resource with the name requested already exists.

Waiters

Waiters poll by repeatedly sending a request until some remote success condition configured by the Wait specification is fulfilled. The Wait specification determines how many attempts should be made, in addition to delay and retry strategies.

ResourceRequestSuccess

newResourceRequestSuccess :: Wait GetResourceRequestStatus

Polls GetResourceRequestStatus every 5 seconds until a successful state is reached. An error is returned after 720 failed checks.

Operations

Some AWS operations return results that are incomplete and require subsequent requests in order to obtain the entire result set. The process of sending subsequent requests to continue where a previous request left off is called pagination. For example, the ListObjects operation of Amazon S3 returns up to 1000 objects at a time, and you must send subsequent requests with the appropriate Marker in order to retrieve the next page of results.

Operations that have an AWSPager instance can transparently perform subsequent requests, correctly setting Markers and other request facets to iterate through the entire result set of a truncated API operation. Operations which support this have an additional note in the documentation.

Many operations have the ability to filter results on the server side. See the individual operation parameters for details.

GetResource

data GetResource

See: newGetResource smart constructor.

Constructors

GetResource' (Maybe Text) (Maybe Text) Text Text

newGetResource

Arguments

:: Text
$sel:typeName:GetResource'

-> Text
$sel:identifier:GetResource'

-> GetResource

Create a value of GetResource with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:typeVersionId:GetResource', getResource_typeVersionId - For private resource types, the type version to use in this resource operation. If you do not specify a resource version, CloudFormation uses the default version.

$sel:roleArn:GetResource', getResource_roleArn - The Amazon Resource Name (ARN) of the Identity and Access Management (IAM) for Cloud Control API to use when performing this resource operation. The role specified must have the permissions required for this operation. The necessary permissions for each event handler are defined in the handlers section of the resource type definition schema.

If you do not specify a role, Cloud Control API uses a temporary session created using your Amazon Web Services user credentials.

For more information, see Specifying credentials in the Amazon Web Services Cloud Control API User Guide.

$sel:typeName:GetResource', getResource_typeName - The name of the resource type.

$sel:identifier:GetResource', getResource_identifier - The identifier for the resource.

You can specify the primary identifier, or any secondary identifier defined for the resource type in its resource schema. You can only specify one identifier. Primary identifiers can be specified as a string or JSON; secondary identifiers must be specified as JSON.

For compound primary identifiers (that is, one that consists of multiple resource properties strung together), to specify the primary identifier as a string, list the property values in the order they are specified in the primary identifier definition, separated by |.

For more information, see Identifying resources in the Amazon Web Services Cloud Control API User Guide.

data GetResourceResponse

See: newGetResourceResponse smart constructor.

Constructors

GetResourceResponse' (Maybe Text) (Maybe ResourceDescription) Int

newGetResourceResponse

Arguments

:: Int
$sel:httpStatus:GetResourceResponse'

-> GetResourceResponse

Create a value of GetResourceResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:typeName:GetResource', getResourceResponse_typeName - The name of the resource type.

$sel:resourceDescription:GetResourceResponse', getResourceResponse_resourceDescription - Undocumented member.

$sel:httpStatus:GetResourceResponse', getResourceResponse_httpStatus - The response's http status code.

ListResourceRequests

data ListResourceRequests

See: newListResourceRequests smart constructor.

Constructors

ListResourceRequests' (Maybe Text) (Maybe ResourceRequestStatusFilter) (Maybe Natural)

newListResourceRequests :: ListResourceRequests

Create a value of ListResourceRequests with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:nextToken:ListResourceRequests', listResourceRequests_nextToken - If the previous paginated request didn't return all of the remaining results, the response object's NextToken parameter value is set to a token. To retrieve the next set of results, call this action again and assign that token to the request object's NextToken parameter. If there are no remaining results, the previous response object's NextToken parameter is set to null.

$sel:resourceRequestStatusFilter:ListResourceRequests', listResourceRequests_resourceRequestStatusFilter - The filter criteria to apply to the requests returned.

$sel:maxResults:ListResourceRequests', listResourceRequests_maxResults - The maximum number of results to be returned with a single call. If the number of available results exceeds this maximum, the response includes a NextToken value that you can assign to the NextToken request parameter to get the next set of results.

The default is 20.

data ListResourceRequestsResponse

See: newListResourceRequestsResponse smart constructor.

Constructors

ListResourceRequestsResponse' (Maybe [ProgressEvent]) (Maybe Text) Int

newListResourceRequestsResponse

Arguments

:: Int
$sel:httpStatus:ListResourceRequestsResponse'

-> ListResourceRequestsResponse

Create a value of ListResourceRequestsResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:resourceRequestStatusSummaries:ListResourceRequestsResponse', listResourceRequestsResponse_resourceRequestStatusSummaries - The requests that match the specified filter criteria.

$sel:nextToken:ListResourceRequests', listResourceRequestsResponse_nextToken - If the request doesn't return all of the remaining results, NextToken is set to a token. To retrieve the next set of results, call ListResources again and assign that token to the request object's NextToken parameter. If the request returns all results, NextToken is set to null.

$sel:httpStatus:ListResourceRequestsResponse', listResourceRequestsResponse_httpStatus - The response's http status code.

CancelResourceRequest

data CancelResourceRequest

See: newCancelResourceRequest smart constructor.

Constructors

CancelResourceRequest' Text

newCancelResourceRequest

Arguments

:: Text
$sel:requestToken:CancelResourceRequest'

-> CancelResourceRequest

Create a value of CancelResourceRequest with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:requestToken:CancelResourceRequest', cancelResourceRequest_requestToken - The RequestToken of the ProgressEvent object returned by the resource operation request.

data CancelResourceRequestResponse

See: newCancelResourceRequestResponse smart constructor.

Constructors

CancelResourceRequestResponse' (Maybe ProgressEvent) Int

newCancelResourceRequestResponse

Arguments

:: Int
$sel:httpStatus:CancelResourceRequestResponse'

-> CancelResourceRequestResponse

Create a value of CancelResourceRequestResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:progressEvent:CancelResourceRequestResponse', cancelResourceRequestResponse_progressEvent - Undocumented member.

$sel:httpStatus:CancelResourceRequestResponse', cancelResourceRequestResponse_httpStatus - The response's http status code.

DeleteResource

data DeleteResource

See: newDeleteResource smart constructor.

Constructors

DeleteResource' (Maybe Text) (Maybe Text) (Maybe Text) Text Text

newDeleteResource

Arguments

:: Text
$sel:typeName:DeleteResource'

-> Text
$sel:identifier:DeleteResource'

-> DeleteResource

Create a value of DeleteResource with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:clientToken:DeleteResource', deleteResource_clientToken - A unique identifier to ensure the idempotency of the resource request. As a best practice, specify this token to ensure idempotency, so that Amazon Web Services Cloud Control API can accurately distinguish between request retries and new resource requests. You might retry a resource request to ensure that it was successfully received.

A client token is valid for 36 hours once used. After that, a resource request with the same client token is treated as a new request.

If you do not specify a client token, one is generated for inclusion in the request.

For more information, see Ensuring resource operation requests are unique in the Amazon Web Services Cloud Control API User Guide.

$sel:typeVersionId:DeleteResource', deleteResource_typeVersionId - For private resource types, the type version to use in this resource operation. If you do not specify a resource version, CloudFormation uses the default version.

$sel:roleArn:DeleteResource', deleteResource_roleArn - The Amazon Resource Name (ARN) of the Identity and Access Management (IAM) for Cloud Control API to use when performing this resource operation. The role specified must have the permissions required for this operation. The necessary permissions for each event handler are defined in the handlers section of the resource type definition schema.

If you do not specify a role, Cloud Control API uses a temporary session created using your Amazon Web Services user credentials.

For more information, see Specifying credentials in the Amazon Web Services Cloud Control API User Guide.

$sel:typeName:DeleteResource', deleteResource_typeName - The name of the resource type.

$sel:identifier:DeleteResource', deleteResource_identifier - The identifier for the resource.

You can specify the primary identifier, or any secondary identifier defined for the resource type in its resource schema. You can only specify one identifier. Primary identifiers can be specified as a string or JSON; secondary identifiers must be specified as JSON.

For compound primary identifiers (that is, one that consists of multiple resource properties strung together), to specify the primary identifier as a string, list the property values in the order they are specified in the primary identifier definition, separated by |.

For more information, see Identifying resources in the Amazon Web Services Cloud Control API User Guide.

data DeleteResourceResponse

See: newDeleteResourceResponse smart constructor.

Constructors

DeleteResourceResponse' (Maybe ProgressEvent) Int

newDeleteResourceResponse

Arguments

:: Int
$sel:httpStatus:DeleteResourceResponse'

-> DeleteResourceResponse

Create a value of DeleteResourceResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:progressEvent:DeleteResourceResponse', deleteResourceResponse_progressEvent - Represents the current status of the resource deletion request.

After you have initiated a resource deletion request, you can monitor the progress of your request by calling GetResourceRequestStatus using the RequestToken of the ProgressEvent returned by DeleteResource.

$sel:httpStatus:DeleteResourceResponse', deleteResourceResponse_httpStatus - The response's http status code.

UpdateResource

data UpdateResource

See: newUpdateResource smart constructor.

Constructors

UpdateResource' (Maybe Text) (Maybe Text) (Maybe Text) Text Text (Sensitive Text)

newUpdateResource

Arguments

:: Text
$sel:typeName:UpdateResource'

-> Text
$sel:identifier:UpdateResource'

-> Text
$sel:patchDocument:UpdateResource'

-> UpdateResource

Create a value of UpdateResource with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:clientToken:UpdateResource', updateResource_clientToken - A unique identifier to ensure the idempotency of the resource request. As a best practice, specify this token to ensure idempotency, so that Amazon Web Services Cloud Control API can accurately distinguish between request retries and new resource requests. You might retry a resource request to ensure that it was successfully received.

A client token is valid for 36 hours once used. After that, a resource request with the same client token is treated as a new request.

If you do not specify a client token, one is generated for inclusion in the request.

For more information, see Ensuring resource operation requests are unique in the Amazon Web Services Cloud Control API User Guide.

$sel:typeVersionId:UpdateResource', updateResource_typeVersionId - For private resource types, the type version to use in this resource operation. If you do not specify a resource version, CloudFormation uses the default version.

$sel:roleArn:UpdateResource', updateResource_roleArn - The Amazon Resource Name (ARN) of the Identity and Access Management (IAM) for Cloud Control API to use when performing this resource operation. The role specified must have the permissions required for this operation. The necessary permissions for each event handler are defined in the handlers section of the resource type definition schema.

If you do not specify a role, Cloud Control API uses a temporary session created using your Amazon Web Services user credentials.

For more information, see Specifying credentials in the Amazon Web Services Cloud Control API User Guide.

$sel:typeName:UpdateResource', updateResource_typeName - The name of the resource type.

$sel:identifier:UpdateResource', updateResource_identifier - The identifier for the resource.

You can specify the primary identifier, or any secondary identifier defined for the resource type in its resource schema. You can only specify one identifier. Primary identifiers can be specified as a string or JSON; secondary identifiers must be specified as JSON.

For compound primary identifiers (that is, one that consists of multiple resource properties strung together), to specify the primary identifier as a string, list the property values in the order they are specified in the primary identifier definition, separated by |.

For more information, see Identifying resources in the Amazon Web Services Cloud Control API User Guide.

$sel:patchDocument:UpdateResource', updateResource_patchDocument - A JavaScript Object Notation (JSON) document listing the patch operations that represent the updates to apply to the current resource properties. For details, see Composing the patch document in the Amazon Web Services Cloud Control API User Guide.

data UpdateResourceResponse

See: newUpdateResourceResponse smart constructor.

Constructors

UpdateResourceResponse' (Maybe ProgressEvent) Int

newUpdateResourceResponse

Arguments

:: Int
$sel:httpStatus:UpdateResourceResponse'

-> UpdateResourceResponse

Create a value of UpdateResourceResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:progressEvent:UpdateResourceResponse', updateResourceResponse_progressEvent - Represents the current status of the resource update request.

Use the RequestToken of the ProgressEvent with GetResourceRequestStatus to return the current status of a resource operation request.

$sel:httpStatus:UpdateResourceResponse', updateResourceResponse_httpStatus - The response's http status code.

ListResources

data ListResources

See: newListResources smart constructor.

Constructors

ListResources' (Maybe (Sensitive Text)) (Maybe Text) (Maybe Text) (Maybe Natural) (Maybe Text) Text

newListResources

Arguments

:: Text
$sel:typeName:ListResources'

-> ListResources

Create a value of ListResources with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:resourceModel:ListResources', listResources_resourceModel - The resource model to use to select the resources to return.

$sel:nextToken:ListResources', listResources_nextToken - If the previous paginated request didn't return all of the remaining results, the response object's NextToken parameter value is set to a token. To retrieve the next set of results, call this action again and assign that token to the request object's NextToken parameter. If there are no remaining results, the previous response object's NextToken parameter is set to null.

$sel:typeVersionId:ListResources', listResources_typeVersionId - For private resource types, the type version to use in this resource operation. If you do not specify a resource version, CloudFormation uses the default version.

$sel:maxResults:ListResources', listResources_maxResults - The maximum number of results to be returned with a single call. If the number of available results exceeds this maximum, the response includes a NextToken value that you can assign to the NextToken request parameter to get the next set of results.

The default is 20.

$sel:roleArn:ListResources', listResources_roleArn - The Amazon Resource Name (ARN) of the Identity and Access Management (IAM) for Cloud Control API to use when performing this resource operation. The role specified must have the permissions required for this operation. The necessary permissions for each event handler are defined in the handlers section of the resource type definition schema.

If you do not specify a role, Cloud Control API uses a temporary session created using your Amazon Web Services user credentials.

For more information, see Specifying credentials in the Amazon Web Services Cloud Control API User Guide.

$sel:typeName:ListResources', listResources_typeName - The name of the resource type.

data ListResourcesResponse

See: newListResourcesResponse smart constructor.

Constructors

ListResourcesResponse' (Maybe [ResourceDescription]) (Maybe Text) (Maybe Text) Int

newListResourcesResponse

Arguments

:: Int
$sel:httpStatus:ListResourcesResponse'

-> ListResourcesResponse

Create a value of ListResourcesResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:resourceDescriptions:ListResourcesResponse', listResourcesResponse_resourceDescriptions - Information about the specified resources, including primary identifier and resource model.

$sel:typeName:ListResources', listResourcesResponse_typeName - The name of the resource type.

$sel:nextToken:ListResources', listResourcesResponse_nextToken - If the request doesn't return all of the remaining results, NextToken is set to a token. To retrieve the next set of results, call ListResources again and assign that token to the request object's NextToken parameter. If the request returns all results, NextToken is set to null.

$sel:httpStatus:ListResourcesResponse', listResourcesResponse_httpStatus - The response's http status code.

CreateResource

data CreateResource

See: newCreateResource smart constructor.

Constructors

CreateResource' (Maybe Text) (Maybe Text) (Maybe Text) Text (Sensitive Text)

newCreateResource

Arguments

:: Text
$sel:typeName:CreateResource'

-> Text
$sel:desiredState:CreateResource'

-> CreateResource

Create a value of CreateResource with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:clientToken:CreateResource', createResource_clientToken - A unique identifier to ensure the idempotency of the resource request. As a best practice, specify this token to ensure idempotency, so that Amazon Web Services Cloud Control API can accurately distinguish between request retries and new resource requests. You might retry a resource request to ensure that it was successfully received.

A client token is valid for 36 hours once used. After that, a resource request with the same client token is treated as a new request.

If you do not specify a client token, one is generated for inclusion in the request.

For more information, see Ensuring resource operation requests are unique in the Amazon Web Services Cloud Control API User Guide.

$sel:typeVersionId:CreateResource', createResource_typeVersionId - For private resource types, the type version to use in this resource operation. If you do not specify a resource version, CloudFormation uses the default version.

$sel:roleArn:CreateResource', createResource_roleArn - The Amazon Resource Name (ARN) of the Identity and Access Management (IAM) for Cloud Control API to use when performing this resource operation. The role specified must have the permissions required for this operation. The necessary permissions for each event handler are defined in the handlers section of the resource type definition schema.

If you do not specify a role, Cloud Control API uses a temporary session created using your Amazon Web Services user credentials.

For more information, see Specifying credentials in the Amazon Web Services Cloud Control API User Guide.

$sel:typeName:CreateResource', createResource_typeName - The name of the resource type.

$sel:desiredState:CreateResource', createResource_desiredState - Structured data format representing the desired state of the resource, consisting of that resource's properties and their desired values.

Cloud Control API currently supports JSON as a structured data format.

Specify the desired state as one of the following:

A JSON blob

A local path containing the desired state in JSON data format

For more information, see Composing the desired state of the resource in the Amazon Web Services Cloud Control API User Guide.

For more information about the properties of a specific resource, refer to the related topic for the resource in the Resource and property types reference in the Amazon Web Services CloudFormation Users Guide.

data CreateResourceResponse

See: newCreateResourceResponse smart constructor.

Constructors

CreateResourceResponse' (Maybe ProgressEvent) Int

newCreateResourceResponse

Arguments

:: Int
$sel:httpStatus:CreateResourceResponse'

-> CreateResourceResponse

Create a value of CreateResourceResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:progressEvent:CreateResourceResponse', createResourceResponse_progressEvent - Represents the current status of the resource creation request.

After you have initiated a resource creation request, you can monitor the progress of your request by calling GetResourceRequestStatus using the RequestToken of the ProgressEvent returned by CreateResource.

$sel:httpStatus:CreateResourceResponse', createResourceResponse_httpStatus - The response's http status code.

GetResourceRequestStatus

data GetResourceRequestStatus

See: newGetResourceRequestStatus smart constructor.

Constructors

GetResourceRequestStatus' Text

newGetResourceRequestStatus

Arguments

:: Text
$sel:requestToken:GetResourceRequestStatus'

-> GetResourceRequestStatus

Create a value of GetResourceRequestStatus with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:requestToken:GetResourceRequestStatus', getResourceRequestStatus_requestToken - A unique token used to track the progress of the resource operation request.

Request tokens are included in the ProgressEvent type returned by a resource operation request.

data GetResourceRequestStatusResponse

See: newGetResourceRequestStatusResponse smart constructor.

Constructors

GetResourceRequestStatusResponse' (Maybe ProgressEvent) Int

newGetResourceRequestStatusResponse

Arguments

:: Int
$sel:httpStatus:GetResourceRequestStatusResponse'

-> GetResourceRequestStatusResponse

Create a value of GetResourceRequestStatusResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:progressEvent:GetResourceRequestStatusResponse', getResourceRequestStatusResponse_progressEvent - Represents the current status of the resource operation request.

$sel:httpStatus:GetResourceRequestStatusResponse', getResourceRequestStatusResponse_httpStatus - The response's http status code.

Types

HandlerErrorCode

newtype HandlerErrorCode

Constructors

HandlerErrorCode'

Fields

fromHandlerErrorCode :: Text

Bundled Patterns

pattern HandlerErrorCode_AccessDenied :: HandlerErrorCode

pattern HandlerErrorCode_AlreadyExists :: HandlerErrorCode

pattern HandlerErrorCode_GeneralServiceException :: HandlerErrorCode

pattern HandlerErrorCode_InternalFailure :: HandlerErrorCode

pattern HandlerErrorCode_InvalidCredentials :: HandlerErrorCode

pattern HandlerErrorCode_InvalidRequest :: HandlerErrorCode

pattern HandlerErrorCode_NetworkFailure :: HandlerErrorCode

pattern HandlerErrorCode_NotFound :: HandlerErrorCode

pattern HandlerErrorCode_NotStabilized :: HandlerErrorCode

pattern HandlerErrorCode_NotUpdatable :: HandlerErrorCode

pattern HandlerErrorCode_ResourceConflict :: HandlerErrorCode

pattern HandlerErrorCode_ServiceInternalError :: HandlerErrorCode

pattern HandlerErrorCode_ServiceLimitExceeded :: HandlerErrorCode

pattern HandlerErrorCode_ServiceTimeout :: HandlerErrorCode

pattern HandlerErrorCode_Throttling :: HandlerErrorCode

Operation

newtype Operation

Constructors

Operation'

Fields

fromOperation :: Text

Bundled Patterns

pattern Operation_CREATE :: Operation

pattern Operation_DELETE :: Operation

pattern Operation_UPDATE :: Operation

OperationStatus

newtype OperationStatus

Constructors

OperationStatus'

Fields

fromOperationStatus :: Text

Bundled Patterns

pattern OperationStatus_CANCEL_COMPLETE :: OperationStatus

pattern OperationStatus_CANCEL_IN_PROGRESS :: OperationStatus

pattern OperationStatus_FAILED :: OperationStatus

pattern OperationStatus_IN_PROGRESS :: OperationStatus

pattern OperationStatus_PENDING :: OperationStatus

pattern OperationStatus_SUCCESS :: OperationStatus

ProgressEvent

data ProgressEvent

Represents the current status of a resource operation request. For more information, see Managing resource operation requests in the Amazon Web Services Cloud Control API User Guide.

See: newProgressEvent smart constructor.

Constructors

ProgressEvent' (Maybe POSIX) (Maybe Text) (Maybe Text) (Maybe (Sensitive Text)) (Maybe Operation) (Maybe Text) (Maybe OperationStatus) (Maybe POSIX) (Maybe Text) (Maybe HandlerErrorCode)

newProgressEvent :: ProgressEvent

Create a value of ProgressEvent with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:retryAfter:ProgressEvent', progressEvent_retryAfter - When to next request the status of this resource operation request.

$sel:typeName:ProgressEvent', progressEvent_typeName - The name of the resource type used in the operation.

$sel:requestToken:ProgressEvent', progressEvent_requestToken - The unique token representing this resource operation request.

Use the RequestToken with GetResourceRequestStatus to return the current status of a resource operation request.

$sel:resourceModel:ProgressEvent', progressEvent_resourceModel - A JSON string containing the resource model, consisting of each resource property and its current value.

$sel:operation:ProgressEvent', progressEvent_operation - The resource operation type.

$sel:identifier:ProgressEvent', progressEvent_identifier - The primary identifier for the resource.

In some cases, the resource identifier may be available before the resource operation has reached a status of SUCCESS.

$sel:operationStatus:ProgressEvent', progressEvent_operationStatus - The current status of the resource operation request.

PENDING: The resource operation has not yet started.

IN_PROGRESS: The resource operation is currently in progress.

SUCCESS: The resource operation has successfully completed.

FAILED: The resource operation has failed. Refer to the error code and status message for more information.

CANCEL_IN_PROGRESS: The resource operation is in the process of being canceled.

CANCEL_COMPLETE: The resource operation has been canceled.

$sel:eventTime:ProgressEvent', progressEvent_eventTime - When the resource operation request was initiated.

$sel:statusMessage:ProgressEvent', progressEvent_statusMessage - Any message explaining the current status.

$sel:errorCode:ProgressEvent', progressEvent_errorCode - For requests with a status of FAILED, the associated error code.

For error code definitions, see Handler error codes in the /CloudFormation Command Line Interface User Guide for Extension Development/.

ResourceDescription

data ResourceDescription

Represents information about a provisioned resource.

See: newResourceDescription smart constructor.

Constructors

ResourceDescription' (Maybe Text) (Maybe (Sensitive Text))

newResourceDescription :: ResourceDescription

Create a value of ResourceDescription with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:identifier:ResourceDescription', resourceDescription_identifier - The primary identifier for the resource.

For more information, see Identifying resources in the Amazon Web Services Cloud Control API User Guide.

$sel:properties:ResourceDescription', resourceDescription_properties - A list of the resource properties and their current values.

ResourceRequestStatusFilter

data ResourceRequestStatusFilter

The filter criteria to use in determining the requests returned.

See: newResourceRequestStatusFilter smart constructor.

Constructors

ResourceRequestStatusFilter' (Maybe [OperationStatus]) (Maybe [Operation])

newResourceRequestStatusFilter :: ResourceRequestStatusFilter

Create a value of ResourceRequestStatusFilter with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:operationStatuses:ResourceRequestStatusFilter', resourceRequestStatusFilter_operationStatuses - The operation statuses to include in the filter.

PENDING: The operation has been requested, but not yet initiated.

IN_PROGRESS: The operation is currently in progress.

SUCCESS: The operation has successfully completed.

FAILED: The operation has failed.

CANCEL_IN_PROGRESS: The operation is currently in the process of being canceled.

CANCEL_COMPLETE: The operation has been canceled.

$sel:operations:ResourceRequestStatusFilter', resourceRequestStatusFilter_operations - The operation types to include in the filter.