Amazonka.CloudControl.CreateResource

Description

Creates the specified resource. For more information, see Creating a resource in the Amazon Web Services Cloud Control API User Guide.

After you have initiated a resource creation request, you can monitor the progress of your request by calling GetResourceRequestStatus using the RequestToken of the ProgressEvent type returned by CreateResource.

Creating a Request

data CreateResource

See: newCreateResource smart constructor.

Constructors

CreateResource'

Fields

clientToken :: Maybe Text

A unique identifier to ensure the idempotency of the resource request. As a best practice, specify this token to ensure idempotency, so that Amazon Web Services Cloud Control API can accurately distinguish between request retries and new resource requests. You might retry a resource request to ensure that it was successfully received.

A client token is valid for 36 hours once used. After that, a resource request with the same client token is treated as a new request.

If you do not specify a client token, one is generated for inclusion in the request.

For more information, see Ensuring resource operation requests are unique in the Amazon Web Services Cloud Control API User Guide.

typeVersionId :: Maybe Text

For private resource types, the type version to use in this resource operation. If you do not specify a resource version, CloudFormation uses the default version.

roleArn :: Maybe Text

The Amazon Resource Name (ARN) of the Identity and Access Management (IAM) for Cloud Control API to use when performing this resource operation. The role specified must have the permissions required for this operation. The necessary permissions for each event handler are defined in the handlers section of the resource type definition schema.

If you do not specify a role, Cloud Control API uses a temporary session created using your Amazon Web Services user credentials.

For more information, see Specifying credentials in the Amazon Web Services Cloud Control API User Guide.

typeName :: Text

The name of the resource type.

desiredState :: Sensitive Text

Structured data format representing the desired state of the resource, consisting of that resource's properties and their desired values.

Cloud Control API currently supports JSON as a structured data format.

Specify the desired state as one of the following:

A JSON blob

A local path containing the desired state in JSON data format

For more information, see Composing the desired state of the resource in the Amazon Web Services Cloud Control API User Guide.

For more information about the properties of a specific resource, refer to the related topic for the resource in the Resource and property types reference in the Amazon Web Services CloudFormation Users Guide.

newCreateResource

Arguments

:: Text
$sel:typeName:CreateResource'

-> Text
$sel:desiredState:CreateResource'

-> CreateResource

Create a value of CreateResource with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:clientToken:CreateResource', createResource_clientToken - A unique identifier to ensure the idempotency of the resource request. As a best practice, specify this token to ensure idempotency, so that Amazon Web Services Cloud Control API can accurately distinguish between request retries and new resource requests. You might retry a resource request to ensure that it was successfully received.

A client token is valid for 36 hours once used. After that, a resource request with the same client token is treated as a new request.

If you do not specify a client token, one is generated for inclusion in the request.

For more information, see Ensuring resource operation requests are unique in the Amazon Web Services Cloud Control API User Guide.

$sel:typeVersionId:CreateResource', createResource_typeVersionId - For private resource types, the type version to use in this resource operation. If you do not specify a resource version, CloudFormation uses the default version.

$sel:roleArn:CreateResource', createResource_roleArn - The Amazon Resource Name (ARN) of the Identity and Access Management (IAM) for Cloud Control API to use when performing this resource operation. The role specified must have the permissions required for this operation. The necessary permissions for each event handler are defined in the handlers section of the resource type definition schema.

If you do not specify a role, Cloud Control API uses a temporary session created using your Amazon Web Services user credentials.

For more information, see Specifying credentials in the Amazon Web Services Cloud Control API User Guide.

$sel:typeName:CreateResource', createResource_typeName - The name of the resource type.

$sel:desiredState:CreateResource', createResource_desiredState - Structured data format representing the desired state of the resource, consisting of that resource's properties and their desired values.

Cloud Control API currently supports JSON as a structured data format.

Specify the desired state as one of the following:

A JSON blob

A local path containing the desired state in JSON data format

For more information, see Composing the desired state of the resource in the Amazon Web Services Cloud Control API User Guide.

For more information about the properties of a specific resource, refer to the related topic for the resource in the Resource and property types reference in the Amazon Web Services CloudFormation Users Guide.

Request Lenses

createResource_clientToken :: Lens' CreateResource (Maybe Text)

A unique identifier to ensure the idempotency of the resource request. As a best practice, specify this token to ensure idempotency, so that Amazon Web Services Cloud Control API can accurately distinguish between request retries and new resource requests. You might retry a resource request to ensure that it was successfully received.

A client token is valid for 36 hours once used. After that, a resource request with the same client token is treated as a new request.

If you do not specify a client token, one is generated for inclusion in the request.

For more information, see Ensuring resource operation requests are unique in the Amazon Web Services Cloud Control API User Guide.

createResource_typeVersionId :: Lens' CreateResource (Maybe Text)

For private resource types, the type version to use in this resource operation. If you do not specify a resource version, CloudFormation uses the default version.

createResource_roleArn :: Lens' CreateResource (Maybe Text)

The Amazon Resource Name (ARN) of the Identity and Access Management (IAM) for Cloud Control API to use when performing this resource operation. The role specified must have the permissions required for this operation. The necessary permissions for each event handler are defined in the handlers section of the resource type definition schema.

If you do not specify a role, Cloud Control API uses a temporary session created using your Amazon Web Services user credentials.

For more information, see Specifying credentials in the Amazon Web Services Cloud Control API User Guide.

createResource_typeName :: Lens' CreateResource Text

The name of the resource type.

createResource_desiredState :: Lens' CreateResource Text

Structured data format representing the desired state of the resource, consisting of that resource's properties and their desired values.

Cloud Control API currently supports JSON as a structured data format.

Specify the desired state as one of the following:

A JSON blob

A local path containing the desired state in JSON data format

For more information, see Composing the desired state of the resource in the Amazon Web Services Cloud Control API User Guide.

For more information about the properties of a specific resource, refer to the related topic for the resource in the Resource and property types reference in the Amazon Web Services CloudFormation Users Guide.

Destructuring the Response

data CreateResourceResponse

See: newCreateResourceResponse smart constructor.

Constructors

CreateResourceResponse'

Fields

progressEvent :: Maybe ProgressEvent

Represents the current status of the resource creation request.

After you have initiated a resource creation request, you can monitor the progress of your request by calling GetResourceRequestStatus using the RequestToken of the ProgressEvent returned by CreateResource.

httpStatus :: Int

The response's http status code.

newCreateResourceResponse

Arguments

:: Int
$sel:httpStatus:CreateResourceResponse'

-> CreateResourceResponse

Create a value of CreateResourceResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:progressEvent:CreateResourceResponse', createResourceResponse_progressEvent - Represents the current status of the resource creation request.

After you have initiated a resource creation request, you can monitor the progress of your request by calling GetResourceRequestStatus using the RequestToken of the ProgressEvent returned by CreateResource.

$sel:httpStatus:CreateResourceResponse', createResourceResponse_httpStatus - The response's http status code.

Response Lenses

createResourceResponse_progressEvent :: Lens' CreateResourceResponse (Maybe ProgressEvent)

Represents the current status of the resource creation request.

After you have initiated a resource creation request, you can monitor the progress of your request by calling GetResourceRequestStatus using the RequestToken of the ProgressEvent returned by CreateResource.

createResourceResponse_httpStatus :: Lens' CreateResourceResponse Int

The response's http status code.