Amazonka.CloudControl.Lens

Description

Operations

GetResource

getResource_typeVersionId :: Lens' GetResource (Maybe Text)

For private resource types, the type version to use in this resource operation. If you do not specify a resource version, CloudFormation uses the default version.

getResource_roleArn :: Lens' GetResource (Maybe Text)

The Amazon Resource Name (ARN) of the Identity and Access Management (IAM) for Cloud Control API to use when performing this resource operation. The role specified must have the permissions required for this operation. The necessary permissions for each event handler are defined in the handlers section of the resource type definition schema.

If you do not specify a role, Cloud Control API uses a temporary session created using your Amazon Web Services user credentials.

For more information, see Specifying credentials in the Amazon Web Services Cloud Control API User Guide.

getResource_typeName :: Lens' GetResource Text

The name of the resource type.

getResource_identifier :: Lens' GetResource Text

The identifier for the resource.

You can specify the primary identifier, or any secondary identifier defined for the resource type in its resource schema. You can only specify one identifier. Primary identifiers can be specified as a string or JSON; secondary identifiers must be specified as JSON.

For compound primary identifiers (that is, one that consists of multiple resource properties strung together), to specify the primary identifier as a string, list the property values in the order they are specified in the primary identifier definition, separated by |.

For more information, see Identifying resources in the Amazon Web Services Cloud Control API User Guide.

getResourceResponse_typeName :: Lens' GetResourceResponse (Maybe Text)

The name of the resource type.

getResourceResponse_resourceDescription :: Lens' GetResourceResponse (Maybe ResourceDescription)

Undocumented member.

getResourceResponse_httpStatus :: Lens' GetResourceResponse Int

The response's http status code.

ListResourceRequests

listResourceRequests_nextToken :: Lens' ListResourceRequests (Maybe Text)

If the previous paginated request didn't return all of the remaining results, the response object's NextToken parameter value is set to a token. To retrieve the next set of results, call this action again and assign that token to the request object's NextToken parameter. If there are no remaining results, the previous response object's NextToken parameter is set to null.

listResourceRequests_resourceRequestStatusFilter :: Lens' ListResourceRequests (Maybe ResourceRequestStatusFilter)

The filter criteria to apply to the requests returned.

listResourceRequests_maxResults :: Lens' ListResourceRequests (Maybe Natural)

The maximum number of results to be returned with a single call. If the number of available results exceeds this maximum, the response includes a NextToken value that you can assign to the NextToken request parameter to get the next set of results.

The default is 20.

listResourceRequestsResponse_resourceRequestStatusSummaries :: Lens' ListResourceRequestsResponse (Maybe [ProgressEvent])

The requests that match the specified filter criteria.

listResourceRequestsResponse_nextToken :: Lens' ListResourceRequestsResponse (Maybe Text)

If the request doesn't return all of the remaining results, NextToken is set to a token. To retrieve the next set of results, call ListResources again and assign that token to the request object's NextToken parameter. If the request returns all results, NextToken is set to null.

listResourceRequestsResponse_httpStatus :: Lens' ListResourceRequestsResponse Int

The response's http status code.

CancelResourceRequest

cancelResourceRequest_requestToken :: Lens' CancelResourceRequest Text

The RequestToken of the ProgressEvent object returned by the resource operation request.

cancelResourceRequestResponse_progressEvent :: Lens' CancelResourceRequestResponse (Maybe ProgressEvent)

Undocumented member.

cancelResourceRequestResponse_httpStatus :: Lens' CancelResourceRequestResponse Int

The response's http status code.

DeleteResource

deleteResource_clientToken :: Lens' DeleteResource (Maybe Text)

A unique identifier to ensure the idempotency of the resource request. As a best practice, specify this token to ensure idempotency, so that Amazon Web Services Cloud Control API can accurately distinguish between request retries and new resource requests. You might retry a resource request to ensure that it was successfully received.

A client token is valid for 36 hours once used. After that, a resource request with the same client token is treated as a new request.

If you do not specify a client token, one is generated for inclusion in the request.

For more information, see Ensuring resource operation requests are unique in the Amazon Web Services Cloud Control API User Guide.

deleteResource_typeVersionId :: Lens' DeleteResource (Maybe Text)

For private resource types, the type version to use in this resource operation. If you do not specify a resource version, CloudFormation uses the default version.

deleteResource_roleArn :: Lens' DeleteResource (Maybe Text)

The Amazon Resource Name (ARN) of the Identity and Access Management (IAM) for Cloud Control API to use when performing this resource operation. The role specified must have the permissions required for this operation. The necessary permissions for each event handler are defined in the handlers section of the resource type definition schema.

If you do not specify a role, Cloud Control API uses a temporary session created using your Amazon Web Services user credentials.

For more information, see Specifying credentials in the Amazon Web Services Cloud Control API User Guide.

deleteResource_typeName :: Lens' DeleteResource Text

The name of the resource type.

deleteResource_identifier :: Lens' DeleteResource Text

The identifier for the resource.

You can specify the primary identifier, or any secondary identifier defined for the resource type in its resource schema. You can only specify one identifier. Primary identifiers can be specified as a string or JSON; secondary identifiers must be specified as JSON.

For compound primary identifiers (that is, one that consists of multiple resource properties strung together), to specify the primary identifier as a string, list the property values in the order they are specified in the primary identifier definition, separated by |.

For more information, see Identifying resources in the Amazon Web Services Cloud Control API User Guide.

deleteResourceResponse_progressEvent :: Lens' DeleteResourceResponse (Maybe ProgressEvent)

Represents the current status of the resource deletion request.

After you have initiated a resource deletion request, you can monitor the progress of your request by calling GetResourceRequestStatus using the RequestToken of the ProgressEvent returned by DeleteResource.

deleteResourceResponse_httpStatus :: Lens' DeleteResourceResponse Int

The response's http status code.

UpdateResource

updateResource_clientToken :: Lens' UpdateResource (Maybe Text)

A unique identifier to ensure the idempotency of the resource request. As a best practice, specify this token to ensure idempotency, so that Amazon Web Services Cloud Control API can accurately distinguish between request retries and new resource requests. You might retry a resource request to ensure that it was successfully received.

A client token is valid for 36 hours once used. After that, a resource request with the same client token is treated as a new request.

If you do not specify a client token, one is generated for inclusion in the request.

For more information, see Ensuring resource operation requests are unique in the Amazon Web Services Cloud Control API User Guide.

updateResource_typeVersionId :: Lens' UpdateResource (Maybe Text)

For private resource types, the type version to use in this resource operation. If you do not specify a resource version, CloudFormation uses the default version.

updateResource_roleArn :: Lens' UpdateResource (Maybe Text)

The Amazon Resource Name (ARN) of the Identity and Access Management (IAM) for Cloud Control API to use when performing this resource operation. The role specified must have the permissions required for this operation. The necessary permissions for each event handler are defined in the handlers section of the resource type definition schema.

If you do not specify a role, Cloud Control API uses a temporary session created using your Amazon Web Services user credentials.

For more information, see Specifying credentials in the Amazon Web Services Cloud Control API User Guide.

updateResource_typeName :: Lens' UpdateResource Text

The name of the resource type.

updateResource_identifier :: Lens' UpdateResource Text

The identifier for the resource.

You can specify the primary identifier, or any secondary identifier defined for the resource type in its resource schema. You can only specify one identifier. Primary identifiers can be specified as a string or JSON; secondary identifiers must be specified as JSON.

For compound primary identifiers (that is, one that consists of multiple resource properties strung together), to specify the primary identifier as a string, list the property values in the order they are specified in the primary identifier definition, separated by |.

For more information, see Identifying resources in the Amazon Web Services Cloud Control API User Guide.

updateResource_patchDocument :: Lens' UpdateResource Text

A JavaScript Object Notation (JSON) document listing the patch operations that represent the updates to apply to the current resource properties. For details, see Composing the patch document in the Amazon Web Services Cloud Control API User Guide.

updateResourceResponse_progressEvent :: Lens' UpdateResourceResponse (Maybe ProgressEvent)

Represents the current status of the resource update request.

Use the RequestToken of the ProgressEvent with GetResourceRequestStatus to return the current status of a resource operation request.

updateResourceResponse_httpStatus :: Lens' UpdateResourceResponse Int

The response's http status code.

ListResources

listResources_resourceModel :: Lens' ListResources (Maybe Text)

The resource model to use to select the resources to return.

listResources_nextToken :: Lens' ListResources (Maybe Text)

If the previous paginated request didn't return all of the remaining results, the response object's NextToken parameter value is set to a token. To retrieve the next set of results, call this action again and assign that token to the request object's NextToken parameter. If there are no remaining results, the previous response object's NextToken parameter is set to null.

listResources_typeVersionId :: Lens' ListResources (Maybe Text)

For private resource types, the type version to use in this resource operation. If you do not specify a resource version, CloudFormation uses the default version.

listResources_maxResults :: Lens' ListResources (Maybe Natural)

The maximum number of results to be returned with a single call. If the number of available results exceeds this maximum, the response includes a NextToken value that you can assign to the NextToken request parameter to get the next set of results.

The default is 20.

listResources_roleArn :: Lens' ListResources (Maybe Text)

The Amazon Resource Name (ARN) of the Identity and Access Management (IAM) for Cloud Control API to use when performing this resource operation. The role specified must have the permissions required for this operation. The necessary permissions for each event handler are defined in the handlers section of the resource type definition schema.

If you do not specify a role, Cloud Control API uses a temporary session created using your Amazon Web Services user credentials.

For more information, see Specifying credentials in the Amazon Web Services Cloud Control API User Guide.

listResources_typeName :: Lens' ListResources Text

The name of the resource type.

listResourcesResponse_resourceDescriptions :: Lens' ListResourcesResponse (Maybe [ResourceDescription])

Information about the specified resources, including primary identifier and resource model.

listResourcesResponse_typeName :: Lens' ListResourcesResponse (Maybe Text)

The name of the resource type.

listResourcesResponse_nextToken :: Lens' ListResourcesResponse (Maybe Text)

If the request doesn't return all of the remaining results, NextToken is set to a token. To retrieve the next set of results, call ListResources again and assign that token to the request object's NextToken parameter. If the request returns all results, NextToken is set to null.

listResourcesResponse_httpStatus :: Lens' ListResourcesResponse Int

The response's http status code.

CreateResource

createResource_clientToken :: Lens' CreateResource (Maybe Text)

A unique identifier to ensure the idempotency of the resource request. As a best practice, specify this token to ensure idempotency, so that Amazon Web Services Cloud Control API can accurately distinguish between request retries and new resource requests. You might retry a resource request to ensure that it was successfully received.

A client token is valid for 36 hours once used. After that, a resource request with the same client token is treated as a new request.

If you do not specify a client token, one is generated for inclusion in the request.

For more information, see Ensuring resource operation requests are unique in the Amazon Web Services Cloud Control API User Guide.

createResource_typeVersionId :: Lens' CreateResource (Maybe Text)

For private resource types, the type version to use in this resource operation. If you do not specify a resource version, CloudFormation uses the default version.

createResource_roleArn :: Lens' CreateResource (Maybe Text)

The Amazon Resource Name (ARN) of the Identity and Access Management (IAM) for Cloud Control API to use when performing this resource operation. The role specified must have the permissions required for this operation. The necessary permissions for each event handler are defined in the handlers section of the resource type definition schema.

If you do not specify a role, Cloud Control API uses a temporary session created using your Amazon Web Services user credentials.

For more information, see Specifying credentials in the Amazon Web Services Cloud Control API User Guide.

createResource_typeName :: Lens' CreateResource Text

The name of the resource type.

createResource_desiredState :: Lens' CreateResource Text

Structured data format representing the desired state of the resource, consisting of that resource's properties and their desired values.

Cloud Control API currently supports JSON as a structured data format.

Specify the desired state as one of the following:

A JSON blob

A local path containing the desired state in JSON data format

For more information, see Composing the desired state of the resource in the Amazon Web Services Cloud Control API User Guide.

For more information about the properties of a specific resource, refer to the related topic for the resource in the Resource and property types reference in the Amazon Web Services CloudFormation Users Guide.

createResourceResponse_progressEvent :: Lens' CreateResourceResponse (Maybe ProgressEvent)

Represents the current status of the resource creation request.

After you have initiated a resource creation request, you can monitor the progress of your request by calling GetResourceRequestStatus using the RequestToken of the ProgressEvent returned by CreateResource.

createResourceResponse_httpStatus :: Lens' CreateResourceResponse Int

The response's http status code.

GetResourceRequestStatus

getResourceRequestStatus_requestToken :: Lens' GetResourceRequestStatus Text

A unique token used to track the progress of the resource operation request.

Request tokens are included in the ProgressEvent type returned by a resource operation request.

getResourceRequestStatusResponse_progressEvent :: Lens' GetResourceRequestStatusResponse (Maybe ProgressEvent)

Represents the current status of the resource operation request.

getResourceRequestStatusResponse_httpStatus :: Lens' GetResourceRequestStatusResponse Int

The response's http status code.

Types

ProgressEvent

progressEvent_retryAfter :: Lens' ProgressEvent (Maybe UTCTime)

When to next request the status of this resource operation request.

progressEvent_typeName :: Lens' ProgressEvent (Maybe Text)

The name of the resource type used in the operation.

progressEvent_requestToken :: Lens' ProgressEvent (Maybe Text)

The unique token representing this resource operation request.

Use the RequestToken with GetResourceRequestStatus to return the current status of a resource operation request.

progressEvent_resourceModel :: Lens' ProgressEvent (Maybe Text)

A JSON string containing the resource model, consisting of each resource property and its current value.

progressEvent_operation :: Lens' ProgressEvent (Maybe Operation)

The resource operation type.

progressEvent_identifier :: Lens' ProgressEvent (Maybe Text)

The primary identifier for the resource.

In some cases, the resource identifier may be available before the resource operation has reached a status of SUCCESS.

progressEvent_operationStatus :: Lens' ProgressEvent (Maybe OperationStatus)

The current status of the resource operation request.

PENDING: The resource operation has not yet started.

IN_PROGRESS: The resource operation is currently in progress.

SUCCESS: The resource operation has successfully completed.

FAILED: The resource operation has failed. Refer to the error code and status message for more information.

CANCEL_IN_PROGRESS: The resource operation is in the process of being canceled.

CANCEL_COMPLETE: The resource operation has been canceled.

progressEvent_eventTime :: Lens' ProgressEvent (Maybe UTCTime)

When the resource operation request was initiated.

progressEvent_statusMessage :: Lens' ProgressEvent (Maybe Text)

Any message explaining the current status.

progressEvent_errorCode :: Lens' ProgressEvent (Maybe HandlerErrorCode)

For requests with a status of FAILED, the associated error code.

For error code definitions, see Handler error codes in the /CloudFormation Command Line Interface User Guide for Extension Development/.

ResourceDescription

resourceDescription_identifier :: Lens' ResourceDescription (Maybe Text)

The primary identifier for the resource.

For more information, see Identifying resources in the Amazon Web Services Cloud Control API User Guide.

resourceDescription_properties :: Lens' ResourceDescription (Maybe Text)

A list of the resource properties and their current values.

ResourceRequestStatusFilter

resourceRequestStatusFilter_operationStatuses :: Lens' ResourceRequestStatusFilter (Maybe [OperationStatus])

The operation statuses to include in the filter.

PENDING: The operation has been requested, but not yet initiated.

IN_PROGRESS: The operation is currently in progress.

SUCCESS: The operation has successfully completed.

FAILED: The operation has failed.

CANCEL_IN_PROGRESS: The operation is currently in the process of being canceled.

CANCEL_COMPLETE: The operation has been canceled.

resourceRequestStatusFilter_operations :: Lens' ResourceRequestStatusFilter (Maybe [Operation])

The operation types to include in the filter.