Amazonka.CloudControl.Types

Description

Service Configuration

defaultService :: Service

API version 2021-09-30 of the Amazon Cloud Control API SDK configuration.

Errors

_TypeNotFoundException :: AsError a => Getting (First ServiceError) a ServiceError

The specified extension does not exist in the CloudFormation registry.

_GeneralServiceException :: AsError a => Getting (First ServiceError) a ServiceError

The resource handler has returned that the downstream service generated an error that does not map to any other handler error code.

_HandlerFailureException :: AsError a => Getting (First ServiceError) a ServiceError

The resource handler has failed without a returning a more specific error code. This can include timeouts.

_InvalidRequestException :: AsError a => Getting (First ServiceError) a ServiceError

The resource handler has returned that invalid input from the user has generated a generic exception.

_PrivateTypeException :: AsError a => Getting (First ServiceError) a ServiceError

Cloud Control API has not received a valid response from the resource handler, due to a configuration error. This includes issues such as the resource handler returning an invalid response, or timing out.

_ServiceInternalErrorException :: AsError a => Getting (First ServiceError) a ServiceError

The resource handler has returned that the downstream service returned an internal error, typically with a 5XX HTTP status code.

_HandlerInternalFailureException :: AsError a => Getting (First ServiceError) a ServiceError

The resource handler has returned that an unexpected error occurred within the resource handler.

_ClientTokenConflictException :: AsError a => Getting (First ServiceError) a ServiceError

The specified client token has already been used in another resource request.

It is best practice for client tokens to be unique for each resource operation request. However, client token expire after 36 hours.

_RequestTokenNotFoundException :: AsError a => Getting (First ServiceError) a ServiceError

A resource operation with the specified request token cannot be found.

_NetworkFailureException :: AsError a => Getting (First ServiceError) a ServiceError

The resource handler has returned that the request could not be completed due to networking issues, such as a failure to receive a response from the server.

_NotStabilizedException :: AsError a => Getting (First ServiceError) a ServiceError

The resource handler has returned that the downstream resource failed to complete all of its ready-state checks.

_NotUpdatableException :: AsError a => Getting (First ServiceError) a ServiceError

One or more properties included in this resource operation are defined as create-only, and therefore cannot be updated.

_UnsupportedActionException :: AsError a => Getting (First ServiceError) a ServiceError

The specified resource does not support this resource operation.

_ServiceLimitExceededException :: AsError a => Getting (First ServiceError) a ServiceError

The resource handler has returned that a non-transient resource limit was reached on the service side.

_ThrottlingException :: AsError a => Getting (First ServiceError) a ServiceError

The request was denied due to request throttling.

_ConcurrentModificationException :: AsError a => Getting (First ServiceError) a ServiceError

The resource is currently being modified by another operation.

_ResourceConflictException :: AsError a => Getting (First ServiceError) a ServiceError

The resource is temporarily unavailable to be acted upon. For example, if the resource is currently undergoing an operation and cannot be acted upon until that operation is finished.

_InvalidCredentialsException :: AsError a => Getting (First ServiceError) a ServiceError

The resource handler has returned that the credentials provided by the user are invalid.

_ResourceNotFoundException :: AsError a => Getting (First ServiceError) a ServiceError

A resource with the specified identifier cannot be found.

_ConcurrentOperationException :: AsError a => Getting (First ServiceError) a ServiceError

Another resource operation is currently being performed on this resource.

_AlreadyExistsException :: AsError a => Getting (First ServiceError) a ServiceError

The resource with the name requested already exists.

HandlerErrorCode

newtype HandlerErrorCode

Constructors

HandlerErrorCode'

Fields

fromHandlerErrorCode :: Text

Bundled Patterns

pattern HandlerErrorCode_AccessDenied :: HandlerErrorCode

pattern HandlerErrorCode_AlreadyExists :: HandlerErrorCode

pattern HandlerErrorCode_GeneralServiceException :: HandlerErrorCode

pattern HandlerErrorCode_InternalFailure :: HandlerErrorCode

pattern HandlerErrorCode_InvalidCredentials :: HandlerErrorCode

pattern HandlerErrorCode_InvalidRequest :: HandlerErrorCode

pattern HandlerErrorCode_NetworkFailure :: HandlerErrorCode

pattern HandlerErrorCode_NotFound :: HandlerErrorCode

pattern HandlerErrorCode_NotStabilized :: HandlerErrorCode

pattern HandlerErrorCode_NotUpdatable :: HandlerErrorCode

pattern HandlerErrorCode_ResourceConflict :: HandlerErrorCode

pattern HandlerErrorCode_ServiceInternalError :: HandlerErrorCode

pattern HandlerErrorCode_ServiceLimitExceeded :: HandlerErrorCode

pattern HandlerErrorCode_ServiceTimeout :: HandlerErrorCode

pattern HandlerErrorCode_Throttling :: HandlerErrorCode

Operation

newtype Operation

Constructors

Operation'

Fields

fromOperation :: Text

Bundled Patterns

pattern Operation_CREATE :: Operation

pattern Operation_DELETE :: Operation

pattern Operation_UPDATE :: Operation

OperationStatus

newtype OperationStatus

Constructors

OperationStatus'

Fields

fromOperationStatus :: Text

Bundled Patterns

pattern OperationStatus_CANCEL_COMPLETE :: OperationStatus

pattern OperationStatus_CANCEL_IN_PROGRESS :: OperationStatus

pattern OperationStatus_FAILED :: OperationStatus

pattern OperationStatus_IN_PROGRESS :: OperationStatus

pattern OperationStatus_PENDING :: OperationStatus

pattern OperationStatus_SUCCESS :: OperationStatus

ProgressEvent

data ProgressEvent

Represents the current status of a resource operation request. For more information, see Managing resource operation requests in the Amazon Web Services Cloud Control API User Guide.

See: newProgressEvent smart constructor.

Constructors

ProgressEvent'

Fields

retryAfter :: Maybe POSIX

When to next request the status of this resource operation request.

typeName :: Maybe Text

The name of the resource type used in the operation.

requestToken :: Maybe Text

The unique token representing this resource operation request.

Use the RequestToken with GetResourceRequestStatus to return the current status of a resource operation request.

resourceModel :: Maybe (Sensitive Text)

A JSON string containing the resource model, consisting of each resource property and its current value.

operation :: Maybe Operation

The resource operation type.

identifier :: Maybe Text

The primary identifier for the resource.

In some cases, the resource identifier may be available before the resource operation has reached a status of SUCCESS.

operationStatus :: Maybe OperationStatus

The current status of the resource operation request.

PENDING: The resource operation has not yet started.

IN_PROGRESS: The resource operation is currently in progress.

SUCCESS: The resource operation has successfully completed.

FAILED: The resource operation has failed. Refer to the error code and status message for more information.

CANCEL_IN_PROGRESS: The resource operation is in the process of being canceled.

CANCEL_COMPLETE: The resource operation has been canceled.

eventTime :: Maybe POSIX

When the resource operation request was initiated.

statusMessage :: Maybe Text

Any message explaining the current status.

errorCode :: Maybe HandlerErrorCode

For requests with a status of FAILED, the associated error code.

For error code definitions, see Handler error codes in the /CloudFormation Command Line Interface User Guide for Extension Development/.

newProgressEvent :: ProgressEvent

Create a value of ProgressEvent with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:retryAfter:ProgressEvent', progressEvent_retryAfter - When to next request the status of this resource operation request.

$sel:typeName:ProgressEvent', progressEvent_typeName - The name of the resource type used in the operation.

$sel:requestToken:ProgressEvent', progressEvent_requestToken - The unique token representing this resource operation request.

Use the RequestToken with GetResourceRequestStatus to return the current status of a resource operation request.

$sel:resourceModel:ProgressEvent', progressEvent_resourceModel - A JSON string containing the resource model, consisting of each resource property and its current value.

$sel:operation:ProgressEvent', progressEvent_operation - The resource operation type.

$sel:identifier:ProgressEvent', progressEvent_identifier - The primary identifier for the resource.

In some cases, the resource identifier may be available before the resource operation has reached a status of SUCCESS.

$sel:operationStatus:ProgressEvent', progressEvent_operationStatus - The current status of the resource operation request.

PENDING: The resource operation has not yet started.

IN_PROGRESS: The resource operation is currently in progress.

SUCCESS: The resource operation has successfully completed.

FAILED: The resource operation has failed. Refer to the error code and status message for more information.

CANCEL_IN_PROGRESS: The resource operation is in the process of being canceled.

CANCEL_COMPLETE: The resource operation has been canceled.

$sel:eventTime:ProgressEvent', progressEvent_eventTime - When the resource operation request was initiated.

$sel:statusMessage:ProgressEvent', progressEvent_statusMessage - Any message explaining the current status.

$sel:errorCode:ProgressEvent', progressEvent_errorCode - For requests with a status of FAILED, the associated error code.

For error code definitions, see Handler error codes in the /CloudFormation Command Line Interface User Guide for Extension Development/.

progressEvent_retryAfter :: Lens' ProgressEvent (Maybe UTCTime)

When to next request the status of this resource operation request.

progressEvent_typeName :: Lens' ProgressEvent (Maybe Text)

The name of the resource type used in the operation.

progressEvent_requestToken :: Lens' ProgressEvent (Maybe Text)

The unique token representing this resource operation request.

Use the RequestToken with GetResourceRequestStatus to return the current status of a resource operation request.

progressEvent_resourceModel :: Lens' ProgressEvent (Maybe Text)

A JSON string containing the resource model, consisting of each resource property and its current value.

progressEvent_operation :: Lens' ProgressEvent (Maybe Operation)

The resource operation type.

progressEvent_identifier :: Lens' ProgressEvent (Maybe Text)

The primary identifier for the resource.

In some cases, the resource identifier may be available before the resource operation has reached a status of SUCCESS.

progressEvent_operationStatus :: Lens' ProgressEvent (Maybe OperationStatus)

The current status of the resource operation request.

PENDING: The resource operation has not yet started.

IN_PROGRESS: The resource operation is currently in progress.

SUCCESS: The resource operation has successfully completed.

FAILED: The resource operation has failed. Refer to the error code and status message for more information.

CANCEL_IN_PROGRESS: The resource operation is in the process of being canceled.

CANCEL_COMPLETE: The resource operation has been canceled.

progressEvent_eventTime :: Lens' ProgressEvent (Maybe UTCTime)

When the resource operation request was initiated.

progressEvent_statusMessage :: Lens' ProgressEvent (Maybe Text)

Any message explaining the current status.

progressEvent_errorCode :: Lens' ProgressEvent (Maybe HandlerErrorCode)

For requests with a status of FAILED, the associated error code.

For error code definitions, see Handler error codes in the /CloudFormation Command Line Interface User Guide for Extension Development/.

ResourceDescription

data ResourceDescription

Represents information about a provisioned resource.

See: newResourceDescription smart constructor.

Constructors

ResourceDescription'

Fields

identifier :: Maybe Text

The primary identifier for the resource.

For more information, see Identifying resources in the Amazon Web Services Cloud Control API User Guide.

properties :: Maybe (Sensitive Text)

A list of the resource properties and their current values.

newResourceDescription :: ResourceDescription

Create a value of ResourceDescription with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:identifier:ResourceDescription', resourceDescription_identifier - The primary identifier for the resource.

For more information, see Identifying resources in the Amazon Web Services Cloud Control API User Guide.

$sel:properties:ResourceDescription', resourceDescription_properties - A list of the resource properties and their current values.

resourceDescription_identifier :: Lens' ResourceDescription (Maybe Text)

The primary identifier for the resource.

For more information, see Identifying resources in the Amazon Web Services Cloud Control API User Guide.

resourceDescription_properties :: Lens' ResourceDescription (Maybe Text)

A list of the resource properties and their current values.

ResourceRequestStatusFilter

data ResourceRequestStatusFilter

The filter criteria to use in determining the requests returned.

See: newResourceRequestStatusFilter smart constructor.

Constructors

ResourceRequestStatusFilter'

Fields

operationStatuses :: Maybe [OperationStatus]

The operation statuses to include in the filter.

PENDING: The operation has been requested, but not yet initiated.

IN_PROGRESS: The operation is currently in progress.

SUCCESS: The operation has successfully completed.

FAILED: The operation has failed.

CANCEL_IN_PROGRESS: The operation is currently in the process of being canceled.

CANCEL_COMPLETE: The operation has been canceled.

operations :: Maybe [Operation]

The operation types to include in the filter.

newResourceRequestStatusFilter :: ResourceRequestStatusFilter

Create a value of ResourceRequestStatusFilter with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:operationStatuses:ResourceRequestStatusFilter', resourceRequestStatusFilter_operationStatuses - The operation statuses to include in the filter.

PENDING: The operation has been requested, but not yet initiated.

IN_PROGRESS: The operation is currently in progress.

SUCCESS: The operation has successfully completed.

FAILED: The operation has failed.

CANCEL_IN_PROGRESS: The operation is currently in the process of being canceled.

CANCEL_COMPLETE: The operation has been canceled.

$sel:operations:ResourceRequestStatusFilter', resourceRequestStatusFilter_operations - The operation types to include in the filter.

resourceRequestStatusFilter_operationStatuses :: Lens' ResourceRequestStatusFilter (Maybe [OperationStatus])

The operation statuses to include in the filter.

PENDING: The operation has been requested, but not yet initiated.

IN_PROGRESS: The operation is currently in progress.

SUCCESS: The operation has successfully completed.

FAILED: The operation has failed.

CANCEL_IN_PROGRESS: The operation is currently in the process of being canceled.

CANCEL_COMPLETE: The operation has been canceled.

resourceRequestStatusFilter_operations :: Lens' ResourceRequestStatusFilter (Maybe [Operation])

The operation types to include in the filter.