Amazonka.CloudControl.Types.ProgressEvent

Description

Documentation

data ProgressEvent

Represents the current status of a resource operation request. For more information, see Managing resource operation requests in the Amazon Web Services Cloud Control API User Guide.

See: newProgressEvent smart constructor.

Constructors

ProgressEvent'

Fields

retryAfter :: Maybe POSIX

When to next request the status of this resource operation request.

typeName :: Maybe Text

The name of the resource type used in the operation.

requestToken :: Maybe Text

The unique token representing this resource operation request.

Use the RequestToken with GetResourceRequestStatus to return the current status of a resource operation request.

resourceModel :: Maybe (Sensitive Text)

A JSON string containing the resource model, consisting of each resource property and its current value.

operation :: Maybe Operation

The resource operation type.

identifier :: Maybe Text

The primary identifier for the resource.

In some cases, the resource identifier may be available before the resource operation has reached a status of SUCCESS.

operationStatus :: Maybe OperationStatus

The current status of the resource operation request.

PENDING: The resource operation has not yet started.

IN_PROGRESS: The resource operation is currently in progress.

SUCCESS: The resource operation has successfully completed.

FAILED: The resource operation has failed. Refer to the error code and status message for more information.

CANCEL_IN_PROGRESS: The resource operation is in the process of being canceled.

CANCEL_COMPLETE: The resource operation has been canceled.

eventTime :: Maybe POSIX

When the resource operation request was initiated.

statusMessage :: Maybe Text

Any message explaining the current status.

errorCode :: Maybe HandlerErrorCode

For requests with a status of FAILED, the associated error code.

For error code definitions, see Handler error codes in the /CloudFormation Command Line Interface User Guide for Extension Development/.

newProgressEvent :: ProgressEvent

Create a value of ProgressEvent with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:retryAfter:ProgressEvent', progressEvent_retryAfter - When to next request the status of this resource operation request.

$sel:typeName:ProgressEvent', progressEvent_typeName - The name of the resource type used in the operation.

$sel:requestToken:ProgressEvent', progressEvent_requestToken - The unique token representing this resource operation request.

Use the RequestToken with GetResourceRequestStatus to return the current status of a resource operation request.

$sel:resourceModel:ProgressEvent', progressEvent_resourceModel - A JSON string containing the resource model, consisting of each resource property and its current value.

$sel:operation:ProgressEvent', progressEvent_operation - The resource operation type.

$sel:identifier:ProgressEvent', progressEvent_identifier - The primary identifier for the resource.

In some cases, the resource identifier may be available before the resource operation has reached a status of SUCCESS.

$sel:operationStatus:ProgressEvent', progressEvent_operationStatus - The current status of the resource operation request.

PENDING: The resource operation has not yet started.

IN_PROGRESS: The resource operation is currently in progress.

SUCCESS: The resource operation has successfully completed.

FAILED: The resource operation has failed. Refer to the error code and status message for more information.

CANCEL_IN_PROGRESS: The resource operation is in the process of being canceled.

CANCEL_COMPLETE: The resource operation has been canceled.

$sel:eventTime:ProgressEvent', progressEvent_eventTime - When the resource operation request was initiated.

$sel:statusMessage:ProgressEvent', progressEvent_statusMessage - Any message explaining the current status.

$sel:errorCode:ProgressEvent', progressEvent_errorCode - For requests with a status of FAILED, the associated error code.

For error code definitions, see Handler error codes in the /CloudFormation Command Line Interface User Guide for Extension Development/.

progressEvent_retryAfter :: Lens' ProgressEvent (Maybe UTCTime)

When to next request the status of this resource operation request.

progressEvent_typeName :: Lens' ProgressEvent (Maybe Text)

The name of the resource type used in the operation.

progressEvent_requestToken :: Lens' ProgressEvent (Maybe Text)

The unique token representing this resource operation request.

Use the RequestToken with GetResourceRequestStatus to return the current status of a resource operation request.

progressEvent_resourceModel :: Lens' ProgressEvent (Maybe Text)

A JSON string containing the resource model, consisting of each resource property and its current value.

progressEvent_operation :: Lens' ProgressEvent (Maybe Operation)

The resource operation type.

progressEvent_identifier :: Lens' ProgressEvent (Maybe Text)

The primary identifier for the resource.

In some cases, the resource identifier may be available before the resource operation has reached a status of SUCCESS.

progressEvent_operationStatus :: Lens' ProgressEvent (Maybe OperationStatus)

The current status of the resource operation request.

PENDING: The resource operation has not yet started.

IN_PROGRESS: The resource operation is currently in progress.

SUCCESS: The resource operation has successfully completed.

FAILED: The resource operation has failed. Refer to the error code and status message for more information.

CANCEL_IN_PROGRESS: The resource operation is in the process of being canceled.

CANCEL_COMPLETE: The resource operation has been canceled.

progressEvent_eventTime :: Lens' ProgressEvent (Maybe UTCTime)

When the resource operation request was initiated.

progressEvent_statusMessage :: Lens' ProgressEvent (Maybe Text)

Any message explaining the current status.

progressEvent_errorCode :: Lens' ProgressEvent (Maybe HandlerErrorCode)

For requests with a status of FAILED, the associated error code.

For error code definitions, see Handler error codes in the /CloudFormation Command Line Interface User Guide for Extension Development/.