Amazonka.CloudControl.Types.ResourceRequestStatusFilter

Description

Documentation

data ResourceRequestStatusFilter

The filter criteria to use in determining the requests returned.

See: newResourceRequestStatusFilter smart constructor.

Constructors

ResourceRequestStatusFilter'

Fields

operationStatuses :: Maybe [OperationStatus]

The operation statuses to include in the filter.

PENDING: The operation has been requested, but not yet initiated.

IN_PROGRESS: The operation is currently in progress.

SUCCESS: The operation has successfully completed.

FAILED: The operation has failed.

CANCEL_IN_PROGRESS: The operation is currently in the process of being canceled.

CANCEL_COMPLETE: The operation has been canceled.

operations :: Maybe [Operation]

The operation types to include in the filter.

newResourceRequestStatusFilter :: ResourceRequestStatusFilter

Create a value of ResourceRequestStatusFilter with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:operationStatuses:ResourceRequestStatusFilter', resourceRequestStatusFilter_operationStatuses - The operation statuses to include in the filter.

PENDING: The operation has been requested, but not yet initiated.

IN_PROGRESS: The operation is currently in progress.

SUCCESS: The operation has successfully completed.

FAILED: The operation has failed.

CANCEL_IN_PROGRESS: The operation is currently in the process of being canceled.

CANCEL_COMPLETE: The operation has been canceled.

$sel:operations:ResourceRequestStatusFilter', resourceRequestStatusFilter_operations - The operation types to include in the filter.

resourceRequestStatusFilter_operationStatuses :: Lens' ResourceRequestStatusFilter (Maybe [OperationStatus])

The operation statuses to include in the filter.

PENDING: The operation has been requested, but not yet initiated.

IN_PROGRESS: The operation is currently in progress.

SUCCESS: The operation has successfully completed.

FAILED: The operation has failed.

CANCEL_IN_PROGRESS: The operation is currently in the process of being canceled.

CANCEL_COMPLETE: The operation has been canceled.

resourceRequestStatusFilter_operations :: Lens' ResourceRequestStatusFilter (Maybe [Operation])

The operation types to include in the filter.