Amazonka.CloudDirectory.BatchWrite

Description

Performs all the write operations in a batch. Either all the operations succeed or none.

Creating a Request

data BatchWrite

See: newBatchWrite smart constructor.

Constructors

BatchWrite'

Fields

directoryArn :: Text

The Amazon Resource Name (ARN) that is associated with the Directory. For more information, see arns.

operations :: [BatchWriteOperation]

A list of operations that are part of the batch.

newBatchWrite

Arguments

:: Text
$sel:directoryArn:BatchWrite'

-> BatchWrite

Create a value of BatchWrite with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:directoryArn:BatchWrite', batchWrite_directoryArn - The Amazon Resource Name (ARN) that is associated with the Directory. For more information, see arns.

$sel:operations:BatchWrite', batchWrite_operations - A list of operations that are part of the batch.

Request Lenses

batchWrite_directoryArn :: Lens' BatchWrite Text

The Amazon Resource Name (ARN) that is associated with the Directory. For more information, see arns.

batchWrite_operations :: Lens' BatchWrite [BatchWriteOperation]

A list of operations that are part of the batch.

Destructuring the Response

data BatchWriteResponse

See: newBatchWriteResponse smart constructor.

Constructors

BatchWriteResponse'

Fields

responses :: Maybe [BatchWriteOperationResponse]

A list of all the responses for each batch write.

httpStatus :: Int

The response's http status code.

newBatchWriteResponse

Arguments

:: Int
$sel:httpStatus:BatchWriteResponse'

-> BatchWriteResponse

Create a value of BatchWriteResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:responses:BatchWriteResponse', batchWriteResponse_responses - A list of all the responses for each batch write.

$sel:httpStatus:BatchWriteResponse', batchWriteResponse_httpStatus - The response's http status code.

Response Lenses

batchWriteResponse_responses :: Lens' BatchWriteResponse (Maybe [BatchWriteOperationResponse])

A list of all the responses for each batch write.

batchWriteResponse_httpStatus :: Lens' BatchWriteResponse Int

The response's http status code.