Amazonka.CloudDirectory.GetAppliedSchemaVersion

Description

Returns current applied schema version ARN, including the minor version in use.

Creating a Request

data GetAppliedSchemaVersion

See: newGetAppliedSchemaVersion smart constructor.

Constructors

GetAppliedSchemaVersion'

Fields

schemaArn :: Text

The ARN of the applied schema.

newGetAppliedSchemaVersion

Arguments

:: Text
$sel:schemaArn:GetAppliedSchemaVersion'

-> GetAppliedSchemaVersion

Create a value of GetAppliedSchemaVersion with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:schemaArn:GetAppliedSchemaVersion', getAppliedSchemaVersion_schemaArn - The ARN of the applied schema.

Request Lenses

getAppliedSchemaVersion_schemaArn :: Lens' GetAppliedSchemaVersion Text

The ARN of the applied schema.

Destructuring the Response

data GetAppliedSchemaVersionResponse

See: newGetAppliedSchemaVersionResponse smart constructor.

Constructors

GetAppliedSchemaVersionResponse'

Fields

appliedSchemaArn :: Maybe Text

Current applied schema ARN, including the minor version in use if one was provided.

httpStatus :: Int

The response's http status code.

newGetAppliedSchemaVersionResponse

Arguments

:: Int
$sel:httpStatus:GetAppliedSchemaVersionResponse'

-> GetAppliedSchemaVersionResponse

Create a value of GetAppliedSchemaVersionResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:appliedSchemaArn:GetAppliedSchemaVersionResponse', getAppliedSchemaVersionResponse_appliedSchemaArn - Current applied schema ARN, including the minor version in use if one was provided.

$sel:httpStatus:GetAppliedSchemaVersionResponse', getAppliedSchemaVersionResponse_httpStatus - The response's http status code.

Response Lenses

getAppliedSchemaVersionResponse_appliedSchemaArn :: Lens' GetAppliedSchemaVersionResponse (Maybe Text)

Current applied schema ARN, including the minor version in use if one was provided.

getAppliedSchemaVersionResponse_httpStatus :: Lens' GetAppliedSchemaVersionResponse Int

The response's http status code.