Amazonka.CloudDirectory.GetLinkAttributes

Description

Retrieves attributes that are associated with a typed link.

Creating a Request

data GetLinkAttributes

See: newGetLinkAttributes smart constructor.

Constructors

GetLinkAttributes'

Fields

consistencyLevel :: Maybe ConsistencyLevel

The consistency level at which to retrieve the attributes on a typed link.

directoryArn :: Text

The Amazon Resource Name (ARN) that is associated with the Directory where the typed link resides. For more information, see arns or Typed Links.

typedLinkSpecifier :: TypedLinkSpecifier

Allows a typed link specifier to be accepted as input.

attributeNames :: [Text]

A list of attribute names whose values will be retrieved.

newGetLinkAttributes

Arguments

:: Text
$sel:directoryArn:GetLinkAttributes'

-> TypedLinkSpecifier
$sel:typedLinkSpecifier:GetLinkAttributes'

-> GetLinkAttributes

Create a value of GetLinkAttributes with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:consistencyLevel:GetLinkAttributes', getLinkAttributes_consistencyLevel - The consistency level at which to retrieve the attributes on a typed link.

$sel:directoryArn:GetLinkAttributes', getLinkAttributes_directoryArn - The Amazon Resource Name (ARN) that is associated with the Directory where the typed link resides. For more information, see arns or Typed Links.

$sel:typedLinkSpecifier:GetLinkAttributes', getLinkAttributes_typedLinkSpecifier - Allows a typed link specifier to be accepted as input.

$sel:attributeNames:GetLinkAttributes', getLinkAttributes_attributeNames - A list of attribute names whose values will be retrieved.

Request Lenses

getLinkAttributes_consistencyLevel :: Lens' GetLinkAttributes (Maybe ConsistencyLevel)

The consistency level at which to retrieve the attributes on a typed link.

getLinkAttributes_directoryArn :: Lens' GetLinkAttributes Text

The Amazon Resource Name (ARN) that is associated with the Directory where the typed link resides. For more information, see arns or Typed Links.

getLinkAttributes_typedLinkSpecifier :: Lens' GetLinkAttributes TypedLinkSpecifier

Allows a typed link specifier to be accepted as input.

getLinkAttributes_attributeNames :: Lens' GetLinkAttributes [Text]

A list of attribute names whose values will be retrieved.

Destructuring the Response

data GetLinkAttributesResponse

See: newGetLinkAttributesResponse smart constructor.

Constructors

GetLinkAttributesResponse'

Fields

attributes :: Maybe [AttributeKeyAndValue]

The attributes that are associated with the typed link.

httpStatus :: Int

The response's http status code.

newGetLinkAttributesResponse

Arguments

:: Int
$sel:httpStatus:GetLinkAttributesResponse'

-> GetLinkAttributesResponse

Create a value of GetLinkAttributesResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:attributes:GetLinkAttributesResponse', getLinkAttributesResponse_attributes - The attributes that are associated with the typed link.

$sel:httpStatus:GetLinkAttributesResponse', getLinkAttributesResponse_httpStatus - The response's http status code.

Response Lenses

getLinkAttributesResponse_attributes :: Lens' GetLinkAttributesResponse (Maybe [AttributeKeyAndValue])

The attributes that are associated with the typed link.

getLinkAttributesResponse_httpStatus :: Lens' GetLinkAttributesResponse Int

The response's http status code.