Amazonka.CloudDirectory.ListAttachedIndices

Description

Lists indices attached to the specified object.

This operation returns paginated results.

Creating a Request

data ListAttachedIndices

See: newListAttachedIndices smart constructor.

Constructors

ListAttachedIndices'

Fields

consistencyLevel :: Maybe ConsistencyLevel

The consistency level to use for this operation.

nextToken :: Maybe Text

The pagination token.

maxResults :: Maybe Natural

The maximum number of results to retrieve.

directoryArn :: Text

The ARN of the directory.

targetReference :: ObjectReference

A reference to the object that has indices attached.

newListAttachedIndices

Arguments

:: Text
$sel:directoryArn:ListAttachedIndices'

-> ObjectReference
$sel:targetReference:ListAttachedIndices'

-> ListAttachedIndices

Create a value of ListAttachedIndices with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:consistencyLevel:ListAttachedIndices', listAttachedIndices_consistencyLevel - The consistency level to use for this operation.

$sel:nextToken:ListAttachedIndices', listAttachedIndices_nextToken - The pagination token.

$sel:maxResults:ListAttachedIndices', listAttachedIndices_maxResults - The maximum number of results to retrieve.

$sel:directoryArn:ListAttachedIndices', listAttachedIndices_directoryArn - The ARN of the directory.

$sel:targetReference:ListAttachedIndices', listAttachedIndices_targetReference - A reference to the object that has indices attached.

Request Lenses

listAttachedIndices_consistencyLevel :: Lens' ListAttachedIndices (Maybe ConsistencyLevel)

The consistency level to use for this operation.

listAttachedIndices_nextToken :: Lens' ListAttachedIndices (Maybe Text)

The pagination token.

listAttachedIndices_maxResults :: Lens' ListAttachedIndices (Maybe Natural)

The maximum number of results to retrieve.

listAttachedIndices_directoryArn :: Lens' ListAttachedIndices Text

The ARN of the directory.

listAttachedIndices_targetReference :: Lens' ListAttachedIndices ObjectReference

A reference to the object that has indices attached.

Destructuring the Response

data ListAttachedIndicesResponse

See: newListAttachedIndicesResponse smart constructor.

Constructors

ListAttachedIndicesResponse'

Fields

indexAttachments :: Maybe [IndexAttachment]

The indices attached to the specified object.

nextToken :: Maybe Text

The pagination token.

httpStatus :: Int

The response's http status code.

newListAttachedIndicesResponse

Arguments

:: Int
$sel:httpStatus:ListAttachedIndicesResponse'

-> ListAttachedIndicesResponse

Create a value of ListAttachedIndicesResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:indexAttachments:ListAttachedIndicesResponse', listAttachedIndicesResponse_indexAttachments - The indices attached to the specified object.

$sel:nextToken:ListAttachedIndices', listAttachedIndicesResponse_nextToken - The pagination token.

$sel:httpStatus:ListAttachedIndicesResponse', listAttachedIndicesResponse_httpStatus - The response's http status code.

Response Lenses

listAttachedIndicesResponse_indexAttachments :: Lens' ListAttachedIndicesResponse (Maybe [IndexAttachment])

The indices attached to the specified object.

listAttachedIndicesResponse_nextToken :: Lens' ListAttachedIndicesResponse (Maybe Text)

The pagination token.

listAttachedIndicesResponse_httpStatus :: Lens' ListAttachedIndicesResponse Int

The response's http status code.