Amazonka.CloudDirectory.ListDirectories

Description

Lists directories created within an account.

This operation returns paginated results.

Creating a Request

data ListDirectories

See: newListDirectories smart constructor.

Constructors

ListDirectories'

Fields

state :: Maybe DirectoryState

The state of the directories in the list. Can be either Enabled, Disabled, or Deleted.

nextToken :: Maybe Text

The pagination token.

maxResults :: Maybe Natural

The maximum number of results to retrieve.

newListDirectories :: ListDirectories

Create a value of ListDirectories with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:state:ListDirectories', listDirectories_state - The state of the directories in the list. Can be either Enabled, Disabled, or Deleted.

$sel:nextToken:ListDirectories', listDirectories_nextToken - The pagination token.

$sel:maxResults:ListDirectories', listDirectories_maxResults - The maximum number of results to retrieve.

Request Lenses

listDirectories_state :: Lens' ListDirectories (Maybe DirectoryState)

The state of the directories in the list. Can be either Enabled, Disabled, or Deleted.

listDirectories_nextToken :: Lens' ListDirectories (Maybe Text)

The pagination token.

listDirectories_maxResults :: Lens' ListDirectories (Maybe Natural)

The maximum number of results to retrieve.

Destructuring the Response

data ListDirectoriesResponse

See: newListDirectoriesResponse smart constructor.

Constructors

ListDirectoriesResponse'

Fields

nextToken :: Maybe Text

The pagination token.

httpStatus :: Int

The response's http status code.

directories :: [Directory]

Lists all directories that are associated with your account in pagination fashion.

newListDirectoriesResponse

Arguments

:: Int
$sel:httpStatus:ListDirectoriesResponse'

-> ListDirectoriesResponse

Create a value of ListDirectoriesResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:nextToken:ListDirectories', listDirectoriesResponse_nextToken - The pagination token.

$sel:httpStatus:ListDirectoriesResponse', listDirectoriesResponse_httpStatus - The response's http status code.

$sel:directories:ListDirectoriesResponse', listDirectoriesResponse_directories - Lists all directories that are associated with your account in pagination fashion.

Response Lenses

listDirectoriesResponse_nextToken :: Lens' ListDirectoriesResponse (Maybe Text)

The pagination token.

listDirectoriesResponse_httpStatus :: Lens' ListDirectoriesResponse Int

The response's http status code.

listDirectoriesResponse_directories :: Lens' ListDirectoriesResponse [Directory]

Lists all directories that are associated with your account in pagination fashion.