Amazonka.CloudDirectory.ListPublishedSchemaArns

Description

Lists the major version families of each published schema. If a major version ARN is provided as SchemaArn, the minor version revisions in that family are listed instead.

This operation returns paginated results.

Creating a Request

data ListPublishedSchemaArns

See: newListPublishedSchemaArns smart constructor.

Constructors

ListPublishedSchemaArns'

Fields

nextToken :: Maybe Text

The pagination token.

schemaArn :: Maybe Text

The response for ListPublishedSchemaArns when this parameter is used will list all minor version ARNs for a major version.

maxResults :: Maybe Natural

The maximum number of results to retrieve.

newListPublishedSchemaArns :: ListPublishedSchemaArns

Create a value of ListPublishedSchemaArns with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:nextToken:ListPublishedSchemaArns', listPublishedSchemaArns_nextToken - The pagination token.

$sel:schemaArn:ListPublishedSchemaArns', listPublishedSchemaArns_schemaArn - The response for ListPublishedSchemaArns when this parameter is used will list all minor version ARNs for a major version.

$sel:maxResults:ListPublishedSchemaArns', listPublishedSchemaArns_maxResults - The maximum number of results to retrieve.

Request Lenses

listPublishedSchemaArns_nextToken :: Lens' ListPublishedSchemaArns (Maybe Text)

The pagination token.

listPublishedSchemaArns_schemaArn :: Lens' ListPublishedSchemaArns (Maybe Text)

The response for ListPublishedSchemaArns when this parameter is used will list all minor version ARNs for a major version.

listPublishedSchemaArns_maxResults :: Lens' ListPublishedSchemaArns (Maybe Natural)

The maximum number of results to retrieve.

Destructuring the Response

data ListPublishedSchemaArnsResponse

See: newListPublishedSchemaArnsResponse smart constructor.

Constructors

ListPublishedSchemaArnsResponse'

Fields

schemaArns :: Maybe [Text]

The ARNs of published schemas.

nextToken :: Maybe Text

The pagination token.

httpStatus :: Int

The response's http status code.

newListPublishedSchemaArnsResponse

Arguments

:: Int
$sel:httpStatus:ListPublishedSchemaArnsResponse'

-> ListPublishedSchemaArnsResponse

Create a value of ListPublishedSchemaArnsResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:schemaArns:ListPublishedSchemaArnsResponse', listPublishedSchemaArnsResponse_schemaArns - The ARNs of published schemas.

$sel:nextToken:ListPublishedSchemaArns', listPublishedSchemaArnsResponse_nextToken - The pagination token.

$sel:httpStatus:ListPublishedSchemaArnsResponse', listPublishedSchemaArnsResponse_httpStatus - The response's http status code.

Response Lenses

listPublishedSchemaArnsResponse_schemaArns :: Lens' ListPublishedSchemaArnsResponse (Maybe [Text])

The ARNs of published schemas.

listPublishedSchemaArnsResponse_nextToken :: Lens' ListPublishedSchemaArnsResponse (Maybe Text)

The pagination token.

listPublishedSchemaArnsResponse_httpStatus :: Lens' ListPublishedSchemaArnsResponse Int

The response's http status code.