Amazonka.CloudDirectory.PutSchemaFromJson

Description

Allows a schema to be updated using JSON upload. Only available for development schemas. See JSON Schema Format for more information.

Creating a Request

data PutSchemaFromJson

See: newPutSchemaFromJson smart constructor.

Constructors

PutSchemaFromJson'

Fields

schemaArn :: Text

The ARN of the schema to update.

document :: Text

The replacement JSON schema.

newPutSchemaFromJson

Arguments

:: Text
$sel:schemaArn:PutSchemaFromJson'

-> Text
$sel:document:PutSchemaFromJson'

-> PutSchemaFromJson

Create a value of PutSchemaFromJson with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:schemaArn:PutSchemaFromJson', putSchemaFromJson_schemaArn - The ARN of the schema to update.

$sel:document:PutSchemaFromJson', putSchemaFromJson_document - The replacement JSON schema.

Request Lenses

putSchemaFromJson_schemaArn :: Lens' PutSchemaFromJson Text

The ARN of the schema to update.

putSchemaFromJson_document :: Lens' PutSchemaFromJson Text

The replacement JSON schema.

Destructuring the Response

data PutSchemaFromJsonResponse

See: newPutSchemaFromJsonResponse smart constructor.

Constructors

PutSchemaFromJsonResponse'

Fields

arn :: Maybe Text

The ARN of the schema to update.

httpStatus :: Int

The response's http status code.

newPutSchemaFromJsonResponse

Arguments

:: Int
$sel:httpStatus:PutSchemaFromJsonResponse'

-> PutSchemaFromJsonResponse

Create a value of PutSchemaFromJsonResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:arn:PutSchemaFromJsonResponse', putSchemaFromJsonResponse_arn - The ARN of the schema to update.

$sel:httpStatus:PutSchemaFromJsonResponse', putSchemaFromJsonResponse_httpStatus - The response's http status code.

Response Lenses

putSchemaFromJsonResponse_arn :: Lens' PutSchemaFromJsonResponse (Maybe Text)

The ARN of the schema to update.

putSchemaFromJsonResponse_httpStatus :: Lens' PutSchemaFromJsonResponse Int

The response's http status code.