Amazonka.CloudDirectory.Types.BatchLookupPolicy

Description

Documentation

data BatchLookupPolicy

Lists all policies from the root of the Directory to the object specified inside a BatchRead operation. For more information, see LookupPolicy and BatchReadRequest$Operations.

See: newBatchLookupPolicy smart constructor.

Constructors

BatchLookupPolicy'

Fields

nextToken :: Maybe Text

The pagination token.

maxResults :: Maybe Natural

The maximum number of results to retrieve.

objectReference :: ObjectReference

Reference that identifies the object whose policies will be looked up.

newBatchLookupPolicy

Arguments

:: ObjectReference
$sel:objectReference:BatchLookupPolicy'

-> BatchLookupPolicy

Create a value of BatchLookupPolicy with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:nextToken:BatchLookupPolicy', batchLookupPolicy_nextToken - The pagination token.

$sel:maxResults:BatchLookupPolicy', batchLookupPolicy_maxResults - The maximum number of results to retrieve.

$sel:objectReference:BatchLookupPolicy', batchLookupPolicy_objectReference - Reference that identifies the object whose policies will be looked up.

batchLookupPolicy_nextToken :: Lens' BatchLookupPolicy (Maybe Text)

The pagination token.

batchLookupPolicy_maxResults :: Lens' BatchLookupPolicy (Maybe Natural)

The maximum number of results to retrieve.

batchLookupPolicy_objectReference :: Lens' BatchLookupPolicy ObjectReference

Reference that identifies the object whose policies will be looked up.