Amazonka.CloudDirectory.Types.BatchReadSuccessfulResponse

Description

Documentation

data BatchReadSuccessfulResponse

Represents the output of a BatchRead success response operation.

See: newBatchReadSuccessfulResponse smart constructor.

Constructors

BatchReadSuccessfulResponse'

Fields

listIndex :: Maybe BatchListIndexResponse

Lists objects attached to the specified index.

getObjectInformation :: Maybe BatchGetObjectInformationResponse

Retrieves metadata about an object.

listAttachedIndices :: Maybe BatchListAttachedIndicesResponse

Lists indices attached to an object.

lookupPolicy :: Maybe BatchLookupPolicyResponse

Lists all policies from the root of the Directory to the object specified. If there are no policies present, an empty list is returned. If policies are present, and if some objects don't have the policies attached, it returns the ObjectIdentifier for such objects. If policies are present, it returns ObjectIdentifier, policyId, and policyType. Paths that don't lead to the root from the target object are ignored. For more information, see Policies.

listObjectParentPaths :: Maybe BatchListObjectParentPathsResponse

Retrieves all available parent paths for any object type such as node, leaf node, policy node, and index node objects. For more information about objects, see Directory Structure.

listObjectAttributes :: Maybe BatchListObjectAttributesResponse

Lists all attributes that are associated with an object.

listIncomingTypedLinks :: Maybe BatchListIncomingTypedLinksResponse

Returns a paginated list of all the incoming TypedLinkSpecifier information for an object. It also supports filtering by typed link facet and identity attributes. For more information, see Typed Links.

getLinkAttributes :: Maybe BatchGetLinkAttributesResponse

The list of attributes to retrieve from the typed link.

getObjectAttributes :: Maybe BatchGetObjectAttributesResponse

Retrieves attributes within a facet that are associated with an object.

listObjectChildren :: Maybe BatchListObjectChildrenResponse

Returns a paginated list of child objects that are associated with a given object.

listObjectParents :: Maybe BatchListObjectParentsResponse

The list of parent objects to retrieve.

listPolicyAttachments :: Maybe BatchListPolicyAttachmentsResponse

Returns all of the ObjectIdentifiers to which a given policy is attached.

listOutgoingTypedLinks :: Maybe BatchListOutgoingTypedLinksResponse

Returns a paginated list of all the outgoing TypedLinkSpecifier information for an object. It also supports filtering by typed link facet and identity attributes. For more information, see Typed Links.

listObjectPolicies :: Maybe BatchListObjectPoliciesResponse

Returns policies attached to an object in pagination fashion.

newBatchReadSuccessfulResponse :: BatchReadSuccessfulResponse

Create a value of BatchReadSuccessfulResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:listIndex:BatchReadSuccessfulResponse', batchReadSuccessfulResponse_listIndex - Lists objects attached to the specified index.

$sel:getObjectInformation:BatchReadSuccessfulResponse', batchReadSuccessfulResponse_getObjectInformation - Retrieves metadata about an object.

$sel:listAttachedIndices:BatchReadSuccessfulResponse', batchReadSuccessfulResponse_listAttachedIndices - Lists indices attached to an object.

$sel:lookupPolicy:BatchReadSuccessfulResponse', batchReadSuccessfulResponse_lookupPolicy - Lists all policies from the root of the Directory to the object specified. If there are no policies present, an empty list is returned. If policies are present, and if some objects don't have the policies attached, it returns the ObjectIdentifier for such objects. If policies are present, it returns ObjectIdentifier, policyId, and policyType. Paths that don't lead to the root from the target object are ignored. For more information, see Policies.

$sel:listObjectParentPaths:BatchReadSuccessfulResponse', batchReadSuccessfulResponse_listObjectParentPaths - Retrieves all available parent paths for any object type such as node, leaf node, policy node, and index node objects. For more information about objects, see Directory Structure.

$sel:listObjectAttributes:BatchReadSuccessfulResponse', batchReadSuccessfulResponse_listObjectAttributes - Lists all attributes that are associated with an object.

$sel:listIncomingTypedLinks:BatchReadSuccessfulResponse', batchReadSuccessfulResponse_listIncomingTypedLinks - Returns a paginated list of all the incoming TypedLinkSpecifier information for an object. It also supports filtering by typed link facet and identity attributes. For more information, see Typed Links.

$sel:getLinkAttributes:BatchReadSuccessfulResponse', batchReadSuccessfulResponse_getLinkAttributes - The list of attributes to retrieve from the typed link.

$sel:getObjectAttributes:BatchReadSuccessfulResponse', batchReadSuccessfulResponse_getObjectAttributes - Retrieves attributes within a facet that are associated with an object.

$sel:listObjectChildren:BatchReadSuccessfulResponse', batchReadSuccessfulResponse_listObjectChildren - Returns a paginated list of child objects that are associated with a given object.

$sel:listObjectParents:BatchReadSuccessfulResponse', batchReadSuccessfulResponse_listObjectParents - The list of parent objects to retrieve.

$sel:listPolicyAttachments:BatchReadSuccessfulResponse', batchReadSuccessfulResponse_listPolicyAttachments - Returns all of the ObjectIdentifiers to which a given policy is attached.

$sel:listOutgoingTypedLinks:BatchReadSuccessfulResponse', batchReadSuccessfulResponse_listOutgoingTypedLinks - Returns a paginated list of all the outgoing TypedLinkSpecifier information for an object. It also supports filtering by typed link facet and identity attributes. For more information, see Typed Links.

$sel:listObjectPolicies:BatchReadSuccessfulResponse', batchReadSuccessfulResponse_listObjectPolicies - Returns policies attached to an object in pagination fashion.

batchReadSuccessfulResponse_listIndex :: Lens' BatchReadSuccessfulResponse (Maybe BatchListIndexResponse)

Lists objects attached to the specified index.

batchReadSuccessfulResponse_getObjectInformation :: Lens' BatchReadSuccessfulResponse (Maybe BatchGetObjectInformationResponse)

Retrieves metadata about an object.

batchReadSuccessfulResponse_listAttachedIndices :: Lens' BatchReadSuccessfulResponse (Maybe BatchListAttachedIndicesResponse)

Lists indices attached to an object.

batchReadSuccessfulResponse_lookupPolicy :: Lens' BatchReadSuccessfulResponse (Maybe BatchLookupPolicyResponse)

Lists all policies from the root of the Directory to the object specified. If there are no policies present, an empty list is returned. If policies are present, and if some objects don't have the policies attached, it returns the ObjectIdentifier for such objects. If policies are present, it returns ObjectIdentifier, policyId, and policyType. Paths that don't lead to the root from the target object are ignored. For more information, see Policies.

batchReadSuccessfulResponse_listObjectParentPaths :: Lens' BatchReadSuccessfulResponse (Maybe BatchListObjectParentPathsResponse)

Retrieves all available parent paths for any object type such as node, leaf node, policy node, and index node objects. For more information about objects, see Directory Structure.

batchReadSuccessfulResponse_listObjectAttributes :: Lens' BatchReadSuccessfulResponse (Maybe BatchListObjectAttributesResponse)

Lists all attributes that are associated with an object.

batchReadSuccessfulResponse_listIncomingTypedLinks :: Lens' BatchReadSuccessfulResponse (Maybe BatchListIncomingTypedLinksResponse)

Returns a paginated list of all the incoming TypedLinkSpecifier information for an object. It also supports filtering by typed link facet and identity attributes. For more information, see Typed Links.

batchReadSuccessfulResponse_getLinkAttributes :: Lens' BatchReadSuccessfulResponse (Maybe BatchGetLinkAttributesResponse)

The list of attributes to retrieve from the typed link.

batchReadSuccessfulResponse_getObjectAttributes :: Lens' BatchReadSuccessfulResponse (Maybe BatchGetObjectAttributesResponse)

Retrieves attributes within a facet that are associated with an object.

batchReadSuccessfulResponse_listObjectChildren :: Lens' BatchReadSuccessfulResponse (Maybe BatchListObjectChildrenResponse)

Returns a paginated list of child objects that are associated with a given object.

batchReadSuccessfulResponse_listObjectParents :: Lens' BatchReadSuccessfulResponse (Maybe BatchListObjectParentsResponse)

The list of parent objects to retrieve.

batchReadSuccessfulResponse_listPolicyAttachments :: Lens' BatchReadSuccessfulResponse (Maybe BatchListPolicyAttachmentsResponse)

Returns all of the ObjectIdentifiers to which a given policy is attached.

batchReadSuccessfulResponse_listOutgoingTypedLinks :: Lens' BatchReadSuccessfulResponse (Maybe BatchListOutgoingTypedLinksResponse)

Returns a paginated list of all the outgoing TypedLinkSpecifier information for an object. It also supports filtering by typed link facet and identity attributes. For more information, see Typed Links.

batchReadSuccessfulResponse_listObjectPolicies :: Lens' BatchReadSuccessfulResponse (Maybe BatchListObjectPoliciesResponse)

Returns policies attached to an object in pagination fashion.