Amazonka.CloudDirectory.Types.BatchWriteOperationResponse

Description

Documentation

data BatchWriteOperationResponse

Represents the output of a BatchWrite response operation.

See: newBatchWriteOperationResponse smart constructor.

Constructors

BatchWriteOperationResponse'

Fields

deleteObject :: Maybe BatchDeleteObjectResponse

Deletes an object in a Directory.

detachFromIndex :: Maybe BatchDetachFromIndexResponse

Detaches the specified object from the specified index.

removeFacetFromObject :: Maybe BatchRemoveFacetFromObjectResponse

The result of a batch remove facet from object operation.

attachObject :: Maybe BatchAttachObjectResponse

Attaches an object to a Directory.

createObject :: Maybe BatchCreateObjectResponse

Creates an object in a Directory.

attachTypedLink :: Maybe BatchAttachTypedLinkResponse

Attaches a typed link to a specified source and target object. For more information, see Typed Links.

detachPolicy :: Maybe BatchDetachPolicyResponse

Detaches a policy from a Directory.

createIndex :: Maybe BatchCreateIndexResponse

Creates an index object. See Indexing and search for more information.

detachObject :: Maybe BatchDetachObjectResponse

Detaches an object from a Directory.

addFacetToObject :: Maybe BatchAddFacetToObjectResponse

The result of an add facet to object batch operation.

detachTypedLink :: Maybe BatchDetachTypedLinkResponse

Detaches a typed link from a specified source and target object. For more information, see Typed Links.

updateObjectAttributes :: Maybe BatchUpdateObjectAttributesResponse

Updates a given object’s attributes.

attachPolicy :: Maybe BatchAttachPolicyResponse

Attaches a policy object to a regular object. An object can have a limited number of attached policies.

updateLinkAttributes :: Maybe BatchUpdateLinkAttributesResponse

Represents the output of a BatchWrite response operation.

attachToIndex :: Maybe BatchAttachToIndexResponse

Attaches the specified object to the specified index.

newBatchWriteOperationResponse :: BatchWriteOperationResponse

Create a value of BatchWriteOperationResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:deleteObject:BatchWriteOperationResponse', batchWriteOperationResponse_deleteObject - Deletes an object in a Directory.

$sel:detachFromIndex:BatchWriteOperationResponse', batchWriteOperationResponse_detachFromIndex - Detaches the specified object from the specified index.

$sel:removeFacetFromObject:BatchWriteOperationResponse', batchWriteOperationResponse_removeFacetFromObject - The result of a batch remove facet from object operation.

$sel:attachObject:BatchWriteOperationResponse', batchWriteOperationResponse_attachObject - Attaches an object to a Directory.

$sel:createObject:BatchWriteOperationResponse', batchWriteOperationResponse_createObject - Creates an object in a Directory.

$sel:attachTypedLink:BatchWriteOperationResponse', batchWriteOperationResponse_attachTypedLink - Attaches a typed link to a specified source and target object. For more information, see Typed Links.

$sel:detachPolicy:BatchWriteOperationResponse', batchWriteOperationResponse_detachPolicy - Detaches a policy from a Directory.

$sel:createIndex:BatchWriteOperationResponse', batchWriteOperationResponse_createIndex - Creates an index object. See Indexing and search for more information.

$sel:detachObject:BatchWriteOperationResponse', batchWriteOperationResponse_detachObject - Detaches an object from a Directory.

$sel:addFacetToObject:BatchWriteOperationResponse', batchWriteOperationResponse_addFacetToObject - The result of an add facet to object batch operation.

$sel:detachTypedLink:BatchWriteOperationResponse', batchWriteOperationResponse_detachTypedLink - Detaches a typed link from a specified source and target object. For more information, see Typed Links.

$sel:updateObjectAttributes:BatchWriteOperationResponse', batchWriteOperationResponse_updateObjectAttributes - Updates a given object’s attributes.

$sel:attachPolicy:BatchWriteOperationResponse', batchWriteOperationResponse_attachPolicy - Attaches a policy object to a regular object. An object can have a limited number of attached policies.

$sel:updateLinkAttributes:BatchWriteOperationResponse', batchWriteOperationResponse_updateLinkAttributes - Represents the output of a BatchWrite response operation.

$sel:attachToIndex:BatchWriteOperationResponse', batchWriteOperationResponse_attachToIndex - Attaches the specified object to the specified index.

batchWriteOperationResponse_deleteObject :: Lens' BatchWriteOperationResponse (Maybe BatchDeleteObjectResponse)

Deletes an object in a Directory.

batchWriteOperationResponse_detachFromIndex :: Lens' BatchWriteOperationResponse (Maybe BatchDetachFromIndexResponse)

Detaches the specified object from the specified index.

batchWriteOperationResponse_removeFacetFromObject :: Lens' BatchWriteOperationResponse (Maybe BatchRemoveFacetFromObjectResponse)

The result of a batch remove facet from object operation.

batchWriteOperationResponse_attachObject :: Lens' BatchWriteOperationResponse (Maybe BatchAttachObjectResponse)

Attaches an object to a Directory.

batchWriteOperationResponse_createObject :: Lens' BatchWriteOperationResponse (Maybe BatchCreateObjectResponse)

Creates an object in a Directory.

batchWriteOperationResponse_attachTypedLink :: Lens' BatchWriteOperationResponse (Maybe BatchAttachTypedLinkResponse)

Attaches a typed link to a specified source and target object. For more information, see Typed Links.

batchWriteOperationResponse_detachPolicy :: Lens' BatchWriteOperationResponse (Maybe BatchDetachPolicyResponse)

Detaches a policy from a Directory.

batchWriteOperationResponse_createIndex :: Lens' BatchWriteOperationResponse (Maybe BatchCreateIndexResponse)

Creates an index object. See Indexing and search for more information.

batchWriteOperationResponse_detachObject :: Lens' BatchWriteOperationResponse (Maybe BatchDetachObjectResponse)

Detaches an object from a Directory.

batchWriteOperationResponse_addFacetToObject :: Lens' BatchWriteOperationResponse (Maybe BatchAddFacetToObjectResponse)

The result of an add facet to object batch operation.

batchWriteOperationResponse_detachTypedLink :: Lens' BatchWriteOperationResponse (Maybe BatchDetachTypedLinkResponse)

Detaches a typed link from a specified source and target object. For more information, see Typed Links.

batchWriteOperationResponse_updateObjectAttributes :: Lens' BatchWriteOperationResponse (Maybe BatchUpdateObjectAttributesResponse)

Updates a given object’s attributes.

batchWriteOperationResponse_attachPolicy :: Lens' BatchWriteOperationResponse (Maybe BatchAttachPolicyResponse)

Attaches a policy object to a regular object. An object can have a limited number of attached policies.

batchWriteOperationResponse_updateLinkAttributes :: Lens' BatchWriteOperationResponse (Maybe BatchUpdateLinkAttributesResponse)

Represents the output of a BatchWrite response operation.

batchWriteOperationResponse_attachToIndex :: Lens' BatchWriteOperationResponse (Maybe BatchAttachToIndexResponse)

Attaches the specified object to the specified index.