Amazonka.CloudDirectory.Types.Directory

Description

Documentation

data Directory

Directory structure that includes the directory name and directory ARN.

See: newDirectory smart constructor.

Constructors

Directory'

Fields

directoryArn :: Maybe Text

The Amazon Resource Name (ARN) that is associated with the directory. For more information, see arns.

state :: Maybe DirectoryState

The state of the directory. Can be either Enabled, Disabled, or Deleted.

name :: Maybe Text

The name of the directory.

creationDateTime :: Maybe POSIX

The date and time when the directory was created.

newDirectory :: Directory

Create a value of Directory with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:directoryArn:Directory', directory_directoryArn - The Amazon Resource Name (ARN) that is associated with the directory. For more information, see arns.

$sel:state:Directory', directory_state - The state of the directory. Can be either Enabled, Disabled, or Deleted.

$sel:name:Directory', directory_name - The name of the directory.

$sel:creationDateTime:Directory', directory_creationDateTime - The date and time when the directory was created.

directory_directoryArn :: Lens' Directory (Maybe Text)

The Amazon Resource Name (ARN) that is associated with the directory. For more information, see arns.

directory_state :: Lens' Directory (Maybe DirectoryState)

The state of the directory. Can be either Enabled, Disabled, or Deleted.

directory_name :: Lens' Directory (Maybe Text)

The name of the directory.

directory_creationDateTime :: Lens' Directory (Maybe UTCTime)

The date and time when the directory was created.