Amazonka.CloudDirectory.Types.PolicyAttachment

Description

Documentation

data PolicyAttachment

Contains the PolicyType, PolicyId, and the ObjectIdentifier to which it is attached. For more information, see Policies.

See: newPolicyAttachment smart constructor.

Constructors

PolicyAttachment'

Fields

policyId :: Maybe Text

The ID of PolicyAttachment.

policyType :: Maybe Text

The type of policy that can be associated with PolicyAttachment.

objectIdentifier :: Maybe Text

The ObjectIdentifier that is associated with PolicyAttachment.

newPolicyAttachment :: PolicyAttachment

Create a value of PolicyAttachment with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:policyId:PolicyAttachment', policyAttachment_policyId - The ID of PolicyAttachment.

$sel:policyType:PolicyAttachment', policyAttachment_policyType - The type of policy that can be associated with PolicyAttachment.

$sel:objectIdentifier:PolicyAttachment', policyAttachment_objectIdentifier - The ObjectIdentifier that is associated with PolicyAttachment.

policyAttachment_policyId :: Lens' PolicyAttachment (Maybe Text)

The ID of PolicyAttachment.

policyAttachment_policyType :: Lens' PolicyAttachment (Maybe Text)

The type of policy that can be associated with PolicyAttachment.

policyAttachment_objectIdentifier :: Lens' PolicyAttachment (Maybe Text)

The ObjectIdentifier that is associated with PolicyAttachment.