Amazonka.CloudDirectory.Types.SchemaFacet

Description

Documentation

data SchemaFacet

A facet.

See: newSchemaFacet smart constructor.

Constructors

SchemaFacet'

Fields

facetName :: Maybe Text

The name of the facet. If this value is set, SchemaArn must also be set.

schemaArn :: Maybe Text

The ARN of the schema that contains the facet with no minor component. See arns and In-Place Schema Upgrade for a description of when to provide minor versions. If this value is set, FacetName must also be set.

newSchemaFacet :: SchemaFacet

Create a value of SchemaFacet with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:facetName:SchemaFacet', schemaFacet_facetName - The name of the facet. If this value is set, SchemaArn must also be set.

$sel:schemaArn:SchemaFacet', schemaFacet_schemaArn - The ARN of the schema that contains the facet with no minor component. See arns and In-Place Schema Upgrade for a description of when to provide minor versions. If this value is set, FacetName must also be set.

schemaFacet_facetName :: Lens' SchemaFacet (Maybe Text)

The name of the facet. If this value is set, SchemaArn must also be set.

schemaFacet_schemaArn :: Lens' SchemaFacet (Maybe Text)

The ARN of the schema that contains the facet with no minor component. See arns and In-Place Schema Upgrade for a description of when to provide minor versions. If this value is set, FacetName must also be set.