Amazonka.CloudDirectory.Types.TypedLinkAttributeDefinition

Description

Documentation

data TypedLinkAttributeDefinition

A typed link attribute definition.

See: newTypedLinkAttributeDefinition smart constructor.

Constructors

TypedLinkAttributeDefinition'

Fields

rules :: Maybe (HashMap Text Rule)

Validation rules that are attached to the attribute definition.

defaultValue :: Maybe TypedAttributeValue

The default value of the attribute (if configured).

isImmutable :: Maybe Bool

Whether the attribute is mutable or not.

name :: Text

The unique name of the typed link attribute.

type' :: FacetAttributeType

The type of the attribute.

requiredBehavior :: RequiredAttributeBehavior

The required behavior of the TypedLinkAttributeDefinition.

newTypedLinkAttributeDefinition

Arguments

:: Text
$sel:name:TypedLinkAttributeDefinition'

-> FacetAttributeType
$sel:type':TypedLinkAttributeDefinition'

-> RequiredAttributeBehavior
$sel:requiredBehavior:TypedLinkAttributeDefinition'

-> TypedLinkAttributeDefinition

Create a value of TypedLinkAttributeDefinition with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:rules:TypedLinkAttributeDefinition', typedLinkAttributeDefinition_rules - Validation rules that are attached to the attribute definition.

$sel:defaultValue:TypedLinkAttributeDefinition', typedLinkAttributeDefinition_defaultValue - The default value of the attribute (if configured).

$sel:isImmutable:TypedLinkAttributeDefinition', typedLinkAttributeDefinition_isImmutable - Whether the attribute is mutable or not.

$sel:name:TypedLinkAttributeDefinition', typedLinkAttributeDefinition_name - The unique name of the typed link attribute.

$sel:type':TypedLinkAttributeDefinition', typedLinkAttributeDefinition_type - The type of the attribute.

$sel:requiredBehavior:TypedLinkAttributeDefinition', typedLinkAttributeDefinition_requiredBehavior - The required behavior of the TypedLinkAttributeDefinition.

typedLinkAttributeDefinition_rules :: Lens' TypedLinkAttributeDefinition (Maybe (HashMap Text Rule))

Validation rules that are attached to the attribute definition.

typedLinkAttributeDefinition_defaultValue :: Lens' TypedLinkAttributeDefinition (Maybe TypedAttributeValue)

The default value of the attribute (if configured).

typedLinkAttributeDefinition_isImmutable :: Lens' TypedLinkAttributeDefinition (Maybe Bool)

Whether the attribute is mutable or not.

typedLinkAttributeDefinition_name :: Lens' TypedLinkAttributeDefinition Text

The unique name of the typed link attribute.

typedLinkAttributeDefinition_type :: Lens' TypedLinkAttributeDefinition FacetAttributeType

The type of the attribute.

typedLinkAttributeDefinition_requiredBehavior :: Lens' TypedLinkAttributeDefinition RequiredAttributeBehavior

The required behavior of the TypedLinkAttributeDefinition.