Amazonka.CloudFormation.DescribeStackDriftDetectionStatus

Description

Returns information about a stack drift detection operation. A stack drift detection operation detects whether a stack's actual configuration differs, or has drifted, from it's expected configuration, as defined in the stack template and any values specified as template parameters. A stack is considered to have drifted if one or more of its resources have drifted. For more information on stack and resource drift, see Detecting Unregulated Configuration Changes to Stacks and Resources.

Use DetectStackDrift to initiate a stack drift detection operation. DetectStackDrift returns a StackDriftDetectionId you can use to monitor the progress of the operation using DescribeStackDriftDetectionStatus. Once the drift detection operation has completed, use DescribeStackResourceDrifts to return drift information about the stack and its resources.

Creating a Request

data DescribeStackDriftDetectionStatus

See: newDescribeStackDriftDetectionStatus smart constructor.

Constructors

DescribeStackDriftDetectionStatus'

Fields

stackDriftDetectionId :: Text

The ID of the drift detection results of this operation.

CloudFormation generates new results, with a new drift detection ID, each time this operation is run. However, the number of drift results CloudFormation retains for any given stack, and for how long, may vary.

newDescribeStackDriftDetectionStatus

Arguments

:: Text
$sel:stackDriftDetectionId:DescribeStackDriftDetectionStatus'

-> DescribeStackDriftDetectionStatus

Create a value of DescribeStackDriftDetectionStatus with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:stackDriftDetectionId:DescribeStackDriftDetectionStatus', describeStackDriftDetectionStatus_stackDriftDetectionId - The ID of the drift detection results of this operation.

CloudFormation generates new results, with a new drift detection ID, each time this operation is run. However, the number of drift results CloudFormation retains for any given stack, and for how long, may vary.

Request Lenses

describeStackDriftDetectionStatus_stackDriftDetectionId :: Lens' DescribeStackDriftDetectionStatus Text

The ID of the drift detection results of this operation.

CloudFormation generates new results, with a new drift detection ID, each time this operation is run. However, the number of drift results CloudFormation retains for any given stack, and for how long, may vary.

Destructuring the Response

data DescribeStackDriftDetectionStatusResponse

See: newDescribeStackDriftDetectionStatusResponse smart constructor.

Constructors

DescribeStackDriftDetectionStatusResponse'

Fields

stackDriftStatus :: Maybe StackDriftStatus

Status of the stack's actual configuration compared to its expected configuration.

DRIFTED: The stack differs from its expected template configuration. A stack is considered to have drifted if one or more of its resources have drifted.

NOT_CHECKED: CloudFormation has not checked if the stack differs from its expected template configuration.

IN_SYNC: The stack's actual configuration matches its expected template configuration.

UNKNOWN: This value is reserved for future use.

driftedStackResourceCount :: Maybe Int

Total number of stack resources that have drifted. This is NULL until the drift detection operation reaches a status of DETECTION_COMPLETE. This value will be 0 for stacks whose drift status is IN_SYNC.

detectionStatusReason :: Maybe Text

The reason the stack drift detection operation has its current status.

httpStatus :: Int

The response's http status code.

stackId :: Text

The ID of the stack.

stackDriftDetectionId :: Text

The ID of the drift detection results of this operation.

CloudFormation generates new results, with a new drift detection ID, each time this operation is run. However, the number of reports CloudFormation retains for any given stack, and for how long, may vary.

detectionStatus :: StackDriftDetectionStatus

The status of the stack drift detection operation.

DETECTION_COMPLETE: The stack drift detection operation has successfully completed for all resources in the stack that support drift detection. (Resources that do not currently support stack detection remain unchecked.)

If you specified logical resource IDs for CloudFormation to use as a filter for the stack drift detection operation, only the resources with those logical IDs are checked for drift.

DETECTION_FAILED: The stack drift detection operation has failed for at least one resource in the stack. Results will be available for resources on which CloudFormation successfully completed drift detection.

DETECTION_IN_PROGRESS: The stack drift detection operation is currently in progress.

timestamp :: ISO8601

Time at which the stack drift detection operation was initiated.

newDescribeStackDriftDetectionStatusResponse

Arguments

:: Int
$sel:httpStatus:DescribeStackDriftDetectionStatusResponse'

-> Text
$sel:stackId:DescribeStackDriftDetectionStatusResponse'

-> Text
$sel:stackDriftDetectionId:DescribeStackDriftDetectionStatus'

-> StackDriftDetectionStatus
$sel:detectionStatus:DescribeStackDriftDetectionStatusResponse'

-> UTCTime
$sel:timestamp:DescribeStackDriftDetectionStatusResponse'

-> DescribeStackDriftDetectionStatusResponse

Create a value of DescribeStackDriftDetectionStatusResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:stackDriftStatus:DescribeStackDriftDetectionStatusResponse', describeStackDriftDetectionStatusResponse_stackDriftStatus - Status of the stack's actual configuration compared to its expected configuration.

DRIFTED: The stack differs from its expected template configuration. A stack is considered to have drifted if one or more of its resources have drifted.

NOT_CHECKED: CloudFormation has not checked if the stack differs from its expected template configuration.

IN_SYNC: The stack's actual configuration matches its expected template configuration.

UNKNOWN: This value is reserved for future use.

$sel:driftedStackResourceCount:DescribeStackDriftDetectionStatusResponse', describeStackDriftDetectionStatusResponse_driftedStackResourceCount - Total number of stack resources that have drifted. This is NULL until the drift detection operation reaches a status of DETECTION_COMPLETE. This value will be 0 for stacks whose drift status is IN_SYNC.

$sel:detectionStatusReason:DescribeStackDriftDetectionStatusResponse', describeStackDriftDetectionStatusResponse_detectionStatusReason - The reason the stack drift detection operation has its current status.

$sel:httpStatus:DescribeStackDriftDetectionStatusResponse', describeStackDriftDetectionStatusResponse_httpStatus - The response's http status code.

$sel:stackId:DescribeStackDriftDetectionStatusResponse', describeStackDriftDetectionStatusResponse_stackId - The ID of the stack.

$sel:stackDriftDetectionId:DescribeStackDriftDetectionStatus', describeStackDriftDetectionStatusResponse_stackDriftDetectionId - The ID of the drift detection results of this operation.

CloudFormation generates new results, with a new drift detection ID, each time this operation is run. However, the number of reports CloudFormation retains for any given stack, and for how long, may vary.

$sel:detectionStatus:DescribeStackDriftDetectionStatusResponse', describeStackDriftDetectionStatusResponse_detectionStatus - The status of the stack drift detection operation.

DETECTION_COMPLETE: The stack drift detection operation has successfully completed for all resources in the stack that support drift detection. (Resources that do not currently support stack detection remain unchecked.)

If you specified logical resource IDs for CloudFormation to use as a filter for the stack drift detection operation, only the resources with those logical IDs are checked for drift.

DETECTION_FAILED: The stack drift detection operation has failed for at least one resource in the stack. Results will be available for resources on which CloudFormation successfully completed drift detection.

DETECTION_IN_PROGRESS: The stack drift detection operation is currently in progress.

$sel:timestamp:DescribeStackDriftDetectionStatusResponse', describeStackDriftDetectionStatusResponse_timestamp - Time at which the stack drift detection operation was initiated.

Response Lenses

describeStackDriftDetectionStatusResponse_stackDriftStatus :: Lens' DescribeStackDriftDetectionStatusResponse (Maybe StackDriftStatus)

Status of the stack's actual configuration compared to its expected configuration.

DRIFTED: The stack differs from its expected template configuration. A stack is considered to have drifted if one or more of its resources have drifted.

NOT_CHECKED: CloudFormation has not checked if the stack differs from its expected template configuration.

IN_SYNC: The stack's actual configuration matches its expected template configuration.

UNKNOWN: This value is reserved for future use.

describeStackDriftDetectionStatusResponse_driftedStackResourceCount :: Lens' DescribeStackDriftDetectionStatusResponse (Maybe Int)

Total number of stack resources that have drifted. This is NULL until the drift detection operation reaches a status of DETECTION_COMPLETE. This value will be 0 for stacks whose drift status is IN_SYNC.

describeStackDriftDetectionStatusResponse_detectionStatusReason :: Lens' DescribeStackDriftDetectionStatusResponse (Maybe Text)

The reason the stack drift detection operation has its current status.

describeStackDriftDetectionStatusResponse_httpStatus :: Lens' DescribeStackDriftDetectionStatusResponse Int

The response's http status code.

describeStackDriftDetectionStatusResponse_stackId :: Lens' DescribeStackDriftDetectionStatusResponse Text

The ID of the stack.

describeStackDriftDetectionStatusResponse_stackDriftDetectionId :: Lens' DescribeStackDriftDetectionStatusResponse Text

The ID of the drift detection results of this operation.

CloudFormation generates new results, with a new drift detection ID, each time this operation is run. However, the number of reports CloudFormation retains for any given stack, and for how long, may vary.

describeStackDriftDetectionStatusResponse_detectionStatus :: Lens' DescribeStackDriftDetectionStatusResponse StackDriftDetectionStatus

The status of the stack drift detection operation.

DETECTION_COMPLETE: The stack drift detection operation has successfully completed for all resources in the stack that support drift detection. (Resources that do not currently support stack detection remain unchecked.)

If you specified logical resource IDs for CloudFormation to use as a filter for the stack drift detection operation, only the resources with those logical IDs are checked for drift.

DETECTION_FAILED: The stack drift detection operation has failed for at least one resource in the stack. Results will be available for resources on which CloudFormation successfully completed drift detection.

DETECTION_IN_PROGRESS: The stack drift detection operation is currently in progress.

describeStackDriftDetectionStatusResponse_timestamp :: Lens' DescribeStackDriftDetectionStatusResponse UTCTime

Time at which the stack drift detection operation was initiated.