Amazonka.CloudFormation.DescribeStackEvents

Description

Returns all stack related events for a specified stack in reverse chronological order. For more information about a stack's event history, go to Stacks in the CloudFormation User Guide.

You can list events for stacks that have failed to create or have been deleted by specifying the unique stack identifier (stack ID).

This operation returns paginated results.

Creating a Request

data DescribeStackEvents

The input for DescribeStackEvents action.

See: newDescribeStackEvents smart constructor.

Constructors

DescribeStackEvents'

Fields

nextToken :: Maybe Text

A string that identifies the next page of events that you want to retrieve.

stackName :: Maybe Text

The name or the unique stack ID that is associated with the stack, which are not always interchangeable:

Running stacks: You can specify either the stack's name or its unique stack ID.

Deleted stacks: You must specify the unique stack ID.

Default: There is no default value.

newDescribeStackEvents :: DescribeStackEvents

Create a value of DescribeStackEvents with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:nextToken:DescribeStackEvents', describeStackEvents_nextToken - A string that identifies the next page of events that you want to retrieve.

$sel:stackName:DescribeStackEvents', describeStackEvents_stackName - The name or the unique stack ID that is associated with the stack, which are not always interchangeable:

Running stacks: You can specify either the stack's name or its unique stack ID.

Deleted stacks: You must specify the unique stack ID.

Default: There is no default value.

Request Lenses

describeStackEvents_nextToken :: Lens' DescribeStackEvents (Maybe Text)

A string that identifies the next page of events that you want to retrieve.

describeStackEvents_stackName :: Lens' DescribeStackEvents (Maybe Text)

The name or the unique stack ID that is associated with the stack, which are not always interchangeable:

Running stacks: You can specify either the stack's name or its unique stack ID.

Deleted stacks: You must specify the unique stack ID.

Default: There is no default value.

Destructuring the Response

data DescribeStackEventsResponse

The output for a DescribeStackEvents action.

See: newDescribeStackEventsResponse smart constructor.

Constructors

DescribeStackEventsResponse'

Fields

nextToken :: Maybe Text

If the output exceeds 1 MB in size, a string that identifies the next page of events. If no additional page exists, this value is null.

stackEvents :: Maybe [StackEvent]

A list of StackEvents structures.

httpStatus :: Int

The response's http status code.

newDescribeStackEventsResponse

Arguments

:: Int
$sel:httpStatus:DescribeStackEventsResponse'

-> DescribeStackEventsResponse

Create a value of DescribeStackEventsResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:nextToken:DescribeStackEvents', describeStackEventsResponse_nextToken - If the output exceeds 1 MB in size, a string that identifies the next page of events. If no additional page exists, this value is null.

$sel:stackEvents:DescribeStackEventsResponse', describeStackEventsResponse_stackEvents - A list of StackEvents structures.

$sel:httpStatus:DescribeStackEventsResponse', describeStackEventsResponse_httpStatus - The response's http status code.

Response Lenses

describeStackEventsResponse_nextToken :: Lens' DescribeStackEventsResponse (Maybe Text)

If the output exceeds 1 MB in size, a string that identifies the next page of events. If no additional page exists, this value is null.

describeStackEventsResponse_stackEvents :: Lens' DescribeStackEventsResponse (Maybe [StackEvent])

A list of StackEvents structures.

describeStackEventsResponse_httpStatus :: Lens' DescribeStackEventsResponse Int

The response's http status code.