Amazonka.CloudFormation.DescribeStackSet

Description

Returns the description of the specified stack set.

Creating a Request

data DescribeStackSet

See: newDescribeStackSet smart constructor.

Constructors

DescribeStackSet'

Fields

callAs :: Maybe CallAs

Service-managed permissions

Specifies whether you are acting as an account administrator in the organization's management account or as a delegated administrator in a member account.

By default, SELF is specified. Use SELF for stack sets with self-managed permissions.

If you are signed in to the management account, specify SELF.

If you are signed in to a delegated administrator account, specify DELEGATED_ADMIN.

Your Amazon Web Services account must be registered as a delegated administrator in the management account. For more information, see Register a delegated administrator in the CloudFormation User Guide.

stackSetName :: Text

The name or unique ID of the stack set whose description you want.

newDescribeStackSet

Arguments

:: Text
$sel:stackSetName:DescribeStackSet'

-> DescribeStackSet

Create a value of DescribeStackSet with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:callAs:DescribeStackSet', describeStackSet_callAs - [Service-managed permissions] Specifies whether you are acting as an account administrator in the organization's management account or as a delegated administrator in a member account.

By default, SELF is specified. Use SELF for stack sets with self-managed permissions.

If you are signed in to the management account, specify SELF.

If you are signed in to a delegated administrator account, specify DELEGATED_ADMIN.

Your Amazon Web Services account must be registered as a delegated administrator in the management account. For more information, see Register a delegated administrator in the CloudFormation User Guide.

$sel:stackSetName:DescribeStackSet', describeStackSet_stackSetName - The name or unique ID of the stack set whose description you want.

Request Lenses

describeStackSet_callAs :: Lens' DescribeStackSet (Maybe CallAs)

Service-managed permissions

Specifies whether you are acting as an account administrator in the organization's management account or as a delegated administrator in a member account.

By default, SELF is specified. Use SELF for stack sets with self-managed permissions.

If you are signed in to the management account, specify SELF.

If you are signed in to a delegated administrator account, specify DELEGATED_ADMIN.

Your Amazon Web Services account must be registered as a delegated administrator in the management account. For more information, see Register a delegated administrator in the CloudFormation User Guide.

describeStackSet_stackSetName :: Lens' DescribeStackSet Text

The name or unique ID of the stack set whose description you want.

Destructuring the Response

data DescribeStackSetResponse

See: newDescribeStackSetResponse smart constructor.

Constructors

DescribeStackSetResponse'

Fields

stackSet :: Maybe StackSet

The specified stack set.

httpStatus :: Int

The response's http status code.

newDescribeStackSetResponse

Arguments

:: Int
$sel:httpStatus:DescribeStackSetResponse'

-> DescribeStackSetResponse

Create a value of DescribeStackSetResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:stackSet:DescribeStackSetResponse', describeStackSetResponse_stackSet - The specified stack set.

$sel:httpStatus:DescribeStackSetResponse', describeStackSetResponse_httpStatus - The response's http status code.

Response Lenses

describeStackSetResponse_stackSet :: Lens' DescribeStackSetResponse (Maybe StackSet)

The specified stack set.

describeStackSetResponse_httpStatus :: Lens' DescribeStackSetResponse Int

The response's http status code.