Amazonka.CloudFormation.DetectStackDrift

Description

Detects whether a stack's actual configuration differs, or has drifted, from it's expected configuration, as defined in the stack template and any values specified as template parameters. For each resource in the stack that supports drift detection, CloudFormation compares the actual configuration of the resource with its expected template configuration. Only resource properties explicitly defined in the stack template are checked for drift. A stack is considered to have drifted if one or more of its resources differ from their expected template configurations. For more information, see Detecting Unregulated Configuration Changes to Stacks and Resources.

Use DetectStackDrift to detect drift on all supported resources for a given stack, or DetectStackResourceDrift to detect drift on individual resources.

For a list of stack resources that currently support drift detection, see Resources that Support Drift Detection.

DetectStackDrift can take up to several minutes, depending on the number of resources contained within the stack. Use DescribeStackDriftDetectionStatus to monitor the progress of a detect stack drift operation. Once the drift detection operation has completed, use DescribeStackResourceDrifts to return drift information about the stack and its resources.

When detecting drift on a stack, CloudFormation does not detect drift on any nested stacks belonging to that stack. Perform DetectStackDrift directly on the nested stack itself.

Creating a Request

data DetectStackDrift

See: newDetectStackDrift smart constructor.

Constructors

DetectStackDrift'

Fields

logicalResourceIds :: Maybe (NonEmpty Text)

The logical names of any resources you want to use as filters.

stackName :: Text

The name of the stack for which you want to detect drift.

newDetectStackDrift

Arguments

:: Text
$sel:stackName:DetectStackDrift'

-> DetectStackDrift

Create a value of DetectStackDrift with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:logicalResourceIds:DetectStackDrift', detectStackDrift_logicalResourceIds - The logical names of any resources you want to use as filters.

$sel:stackName:DetectStackDrift', detectStackDrift_stackName - The name of the stack for which you want to detect drift.

Request Lenses

detectStackDrift_logicalResourceIds :: Lens' DetectStackDrift (Maybe (NonEmpty Text))

The logical names of any resources you want to use as filters.

detectStackDrift_stackName :: Lens' DetectStackDrift Text

The name of the stack for which you want to detect drift.

Destructuring the Response

data DetectStackDriftResponse

See: newDetectStackDriftResponse smart constructor.

Constructors

DetectStackDriftResponse'

Fields

httpStatus :: Int

The response's http status code.

stackDriftDetectionId :: Text

The ID of the drift detection results of this operation.

CloudFormation generates new results, with a new drift detection ID, each time this operation is run. However, the number of drift results CloudFormation retains for any given stack, and for how long, may vary.

newDetectStackDriftResponse

Arguments

:: Int
$sel:httpStatus:DetectStackDriftResponse'

-> Text
$sel:stackDriftDetectionId:DetectStackDriftResponse'

-> DetectStackDriftResponse

Create a value of DetectStackDriftResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:httpStatus:DetectStackDriftResponse', detectStackDriftResponse_httpStatus - The response's http status code.

$sel:stackDriftDetectionId:DetectStackDriftResponse', detectStackDriftResponse_stackDriftDetectionId - The ID of the drift detection results of this operation.

CloudFormation generates new results, with a new drift detection ID, each time this operation is run. However, the number of drift results CloudFormation retains for any given stack, and for how long, may vary.

Response Lenses

detectStackDriftResponse_httpStatus :: Lens' DetectStackDriftResponse Int

The response's http status code.

detectStackDriftResponse_stackDriftDetectionId :: Lens' DetectStackDriftResponse Text

The ID of the drift detection results of this operation.

CloudFormation generates new results, with a new drift detection ID, each time this operation is run. However, the number of drift results CloudFormation retains for any given stack, and for how long, may vary.