Amazonka.CloudFormation.DetectStackResourceDrift

Description

Returns information about whether a resource's actual configuration differs, or has drifted, from it's expected configuration, as defined in the stack template and any values specified as template parameters. This information includes actual and expected property values for resources in which CloudFormation detects drift. Only resource properties explicitly defined in the stack template are checked for drift. For more information about stack and resource drift, see Detecting Unregulated Configuration Changes to Stacks and Resources.

Use DetectStackResourceDrift to detect drift on individual resources, or DetectStackDrift to detect drift on all resources in a given stack that support drift detection.

Resources that do not currently support drift detection cannot be checked. For a list of resources that support drift detection, see Resources that Support Drift Detection.

Creating a Request

data DetectStackResourceDrift

See: newDetectStackResourceDrift smart constructor.

Constructors

DetectStackResourceDrift'

Fields

stackName :: Text

The name of the stack to which the resource belongs.

logicalResourceId :: Text

The logical name of the resource for which to return drift information.

newDetectStackResourceDrift

Arguments

:: Text
$sel:stackName:DetectStackResourceDrift'

-> Text
$sel:logicalResourceId:DetectStackResourceDrift'

-> DetectStackResourceDrift

Create a value of DetectStackResourceDrift with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:stackName:DetectStackResourceDrift', detectStackResourceDrift_stackName - The name of the stack to which the resource belongs.

$sel:logicalResourceId:DetectStackResourceDrift', detectStackResourceDrift_logicalResourceId - The logical name of the resource for which to return drift information.

Request Lenses

detectStackResourceDrift_stackName :: Lens' DetectStackResourceDrift Text

The name of the stack to which the resource belongs.

detectStackResourceDrift_logicalResourceId :: Lens' DetectStackResourceDrift Text

The logical name of the resource for which to return drift information.

Destructuring the Response

data DetectStackResourceDriftResponse

See: newDetectStackResourceDriftResponse smart constructor.

Constructors

DetectStackResourceDriftResponse'

Fields

httpStatus :: Int

The response's http status code.

stackResourceDrift :: StackResourceDrift

Information about whether the resource's actual configuration has drifted from its expected template configuration, including actual and expected property values and any differences detected.

newDetectStackResourceDriftResponse

Arguments

:: Int
$sel:httpStatus:DetectStackResourceDriftResponse'

-> StackResourceDrift
$sel:stackResourceDrift:DetectStackResourceDriftResponse'

-> DetectStackResourceDriftResponse

Create a value of DetectStackResourceDriftResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:httpStatus:DetectStackResourceDriftResponse', detectStackResourceDriftResponse_httpStatus - The response's http status code.

$sel:stackResourceDrift:DetectStackResourceDriftResponse', detectStackResourceDriftResponse_stackResourceDrift - Information about whether the resource's actual configuration has drifted from its expected template configuration, including actual and expected property values and any differences detected.

Response Lenses

detectStackResourceDriftResponse_httpStatus :: Lens' DetectStackResourceDriftResponse Int

The response's http status code.

detectStackResourceDriftResponse_stackResourceDrift :: Lens' DetectStackResourceDriftResponse StackResourceDrift

Information about whether the resource's actual configuration has drifted from its expected template configuration, including actual and expected property values and any differences detected.