Amazonka.CloudFormation.ExecuteChangeSet

Description

Updates a stack using the input information that was provided when the specified change set was created. After the call successfully completes, CloudFormation starts updating the stack. Use the DescribeStacks action to view the status of the update.

When you execute a change set, CloudFormation deletes all other change sets associated with the stack because they aren't valid for the updated stack.

If a stack policy is associated with the stack, CloudFormation enforces the policy during the update. You can't specify a temporary stack policy that overrides the current policy.

To create a change set for the entire stack hierachy, IncludeNestedStacks must have been set to True.

Creating a Request

data ExecuteChangeSet

The input for the ExecuteChangeSet action.

See: newExecuteChangeSet smart constructor.

Constructors

ExecuteChangeSet'

Fields

disableRollback :: Maybe Bool

Preserves the state of previously provisioned resources when an operation fails.

Default: True

clientRequestToken :: Maybe Text

A unique identifier for this ExecuteChangeSet request. Specify this token if you plan to retry requests so that CloudFormation knows that you're not attempting to execute a change set to update a stack with the same name. You might retry ExecuteChangeSet requests to ensure that CloudFormation successfully received them.

stackName :: Maybe Text

If you specified the name of a change set, specify the stack name or ID (ARN) that is associated with the change set you want to execute.

changeSetName :: Text

The name or ARN of the change set that you want use to update the specified stack.

newExecuteChangeSet

Arguments

:: Text
$sel:changeSetName:ExecuteChangeSet'

-> ExecuteChangeSet

Create a value of ExecuteChangeSet with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:disableRollback:ExecuteChangeSet', executeChangeSet_disableRollback - Preserves the state of previously provisioned resources when an operation fails.

Default: True

$sel:clientRequestToken:ExecuteChangeSet', executeChangeSet_clientRequestToken - A unique identifier for this ExecuteChangeSet request. Specify this token if you plan to retry requests so that CloudFormation knows that you're not attempting to execute a change set to update a stack with the same name. You might retry ExecuteChangeSet requests to ensure that CloudFormation successfully received them.

$sel:stackName:ExecuteChangeSet', executeChangeSet_stackName - If you specified the name of a change set, specify the stack name or ID (ARN) that is associated with the change set you want to execute.

$sel:changeSetName:ExecuteChangeSet', executeChangeSet_changeSetName - The name or ARN of the change set that you want use to update the specified stack.

Request Lenses

executeChangeSet_disableRollback :: Lens' ExecuteChangeSet (Maybe Bool)

Preserves the state of previously provisioned resources when an operation fails.

Default: True

executeChangeSet_clientRequestToken :: Lens' ExecuteChangeSet (Maybe Text)

A unique identifier for this ExecuteChangeSet request. Specify this token if you plan to retry requests so that CloudFormation knows that you're not attempting to execute a change set to update a stack with the same name. You might retry ExecuteChangeSet requests to ensure that CloudFormation successfully received them.

executeChangeSet_stackName :: Lens' ExecuteChangeSet (Maybe Text)

If you specified the name of a change set, specify the stack name or ID (ARN) that is associated with the change set you want to execute.

executeChangeSet_changeSetName :: Lens' ExecuteChangeSet Text

The name or ARN of the change set that you want use to update the specified stack.

Destructuring the Response

data ExecuteChangeSetResponse

The output for the ExecuteChangeSet action.

See: newExecuteChangeSetResponse smart constructor.

Constructors

ExecuteChangeSetResponse'

Fields

httpStatus :: Int

The response's http status code.

newExecuteChangeSetResponse

Arguments

:: Int
$sel:httpStatus:ExecuteChangeSetResponse'

-> ExecuteChangeSetResponse

Create a value of ExecuteChangeSetResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:httpStatus:ExecuteChangeSetResponse', executeChangeSetResponse_httpStatus - The response's http status code.

Response Lenses

executeChangeSetResponse_httpStatus :: Lens' ExecuteChangeSetResponse Int

The response's http status code.