Amazonka.CloudFormation.ListImports

Description

Lists all stacks that are importing an exported output value. To modify or remove an exported output value, first use this action to see which stacks are using it. To see the exported output values in your account, see ListExports.

For more information about importing an exported output value, see the Fn::ImportValue function.

This operation returns paginated results.

Creating a Request

data ListImports

See: newListImports smart constructor.

Constructors

ListImports'

Fields

nextToken :: Maybe Text

A string (provided by the ListImports response output) that identifies the next page of stacks that are importing the specified exported output value.

exportName :: Text

The name of the exported output value. CloudFormation returns the stack names that are importing this value.

newListImports

Arguments

:: Text
$sel:exportName:ListImports'

-> ListImports

Create a value of ListImports with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:nextToken:ListImports', listImports_nextToken - A string (provided by the ListImports response output) that identifies the next page of stacks that are importing the specified exported output value.

$sel:exportName:ListImports', listImports_exportName - The name of the exported output value. CloudFormation returns the stack names that are importing this value.

Request Lenses

listImports_nextToken :: Lens' ListImports (Maybe Text)

A string (provided by the ListImports response output) that identifies the next page of stacks that are importing the specified exported output value.

listImports_exportName :: Lens' ListImports Text

The name of the exported output value. CloudFormation returns the stack names that are importing this value.

Destructuring the Response

data ListImportsResponse

See: newListImportsResponse smart constructor.

Constructors

ListImportsResponse'

Fields

imports :: Maybe [Text]

A list of stack names that are importing the specified exported output value.

nextToken :: Maybe Text

A string that identifies the next page of exports. If there is no additional page, this value is null.

httpStatus :: Int

The response's http status code.

newListImportsResponse

Arguments

:: Int
$sel:httpStatus:ListImportsResponse'

-> ListImportsResponse

Create a value of ListImportsResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:imports:ListImportsResponse', listImportsResponse_imports - A list of stack names that are importing the specified exported output value.

$sel:nextToken:ListImports', listImportsResponse_nextToken - A string that identifies the next page of exports. If there is no additional page, this value is null.

$sel:httpStatus:ListImportsResponse', listImportsResponse_httpStatus - The response's http status code.

Response Lenses

listImportsResponse_imports :: Lens' ListImportsResponse (Maybe [Text])

A list of stack names that are importing the specified exported output value.

listImportsResponse_nextToken :: Lens' ListImportsResponse (Maybe Text)

A string that identifies the next page of exports. If there is no additional page, this value is null.

listImportsResponse_httpStatus :: Lens' ListImportsResponse Int

The response's http status code.