Amazonka.CloudFormation.ListTypeRegistrations

Description

Returns a list of registration tokens for the specified extension(s).

Creating a Request

data ListTypeRegistrations

See: newListTypeRegistrations smart constructor.

Constructors

ListTypeRegistrations'

Fields

typeName :: Maybe Text

The name of the extension.

Conditional: You must specify either TypeName and Type, or Arn.

registrationStatusFilter :: Maybe RegistrationStatus

The current status of the extension registration request.

The default is IN_PROGRESS.

nextToken :: Maybe Text

If the previous paginated request didn't return all of the remaining results, the response object's NextToken parameter value is set to a token. To retrieve the next set of results, call this action again and assign that token to the request object's NextToken parameter. If there are no remaining results, the previous response object's NextToken parameter is set to null.

typeArn :: Maybe Text

The Amazon Resource Name (ARN) of the extension.

Conditional: You must specify either TypeName and Type, or Arn.

type' :: Maybe RegistryType

The kind of extension.

Conditional: You must specify either TypeName and Type, or Arn.

maxResults :: Maybe Natural

The maximum number of results to be returned with a single call. If the number of available results exceeds this maximum, the response includes a NextToken value that you can assign to the NextToken request parameter to get the next set of results.

newListTypeRegistrations :: ListTypeRegistrations

Create a value of ListTypeRegistrations with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:typeName:ListTypeRegistrations', listTypeRegistrations_typeName - The name of the extension.

Conditional: You must specify either TypeName and Type, or Arn.

$sel:registrationStatusFilter:ListTypeRegistrations', listTypeRegistrations_registrationStatusFilter - The current status of the extension registration request.

The default is IN_PROGRESS.

$sel:nextToken:ListTypeRegistrations', listTypeRegistrations_nextToken - If the previous paginated request didn't return all of the remaining results, the response object's NextToken parameter value is set to a token. To retrieve the next set of results, call this action again and assign that token to the request object's NextToken parameter. If there are no remaining results, the previous response object's NextToken parameter is set to null.

$sel:typeArn:ListTypeRegistrations', listTypeRegistrations_typeArn - The Amazon Resource Name (ARN) of the extension.

Conditional: You must specify either TypeName and Type, or Arn.

$sel:type':ListTypeRegistrations', listTypeRegistrations_type - The kind of extension.

Conditional: You must specify either TypeName and Type, or Arn.

$sel:maxResults:ListTypeRegistrations', listTypeRegistrations_maxResults - The maximum number of results to be returned with a single call. If the number of available results exceeds this maximum, the response includes a NextToken value that you can assign to the NextToken request parameter to get the next set of results.

Request Lenses

listTypeRegistrations_typeName :: Lens' ListTypeRegistrations (Maybe Text)

The name of the extension.

Conditional: You must specify either TypeName and Type, or Arn.

listTypeRegistrations_registrationStatusFilter :: Lens' ListTypeRegistrations (Maybe RegistrationStatus)

The current status of the extension registration request.

The default is IN_PROGRESS.

listTypeRegistrations_nextToken :: Lens' ListTypeRegistrations (Maybe Text)

If the previous paginated request didn't return all of the remaining results, the response object's NextToken parameter value is set to a token. To retrieve the next set of results, call this action again and assign that token to the request object's NextToken parameter. If there are no remaining results, the previous response object's NextToken parameter is set to null.

listTypeRegistrations_typeArn :: Lens' ListTypeRegistrations (Maybe Text)

The Amazon Resource Name (ARN) of the extension.

Conditional: You must specify either TypeName and Type, or Arn.

listTypeRegistrations_type :: Lens' ListTypeRegistrations (Maybe RegistryType)

The kind of extension.

Conditional: You must specify either TypeName and Type, or Arn.

listTypeRegistrations_maxResults :: Lens' ListTypeRegistrations (Maybe Natural)

The maximum number of results to be returned with a single call. If the number of available results exceeds this maximum, the response includes a NextToken value that you can assign to the NextToken request parameter to get the next set of results.

Destructuring the Response

data ListTypeRegistrationsResponse

See: newListTypeRegistrationsResponse smart constructor.

Constructors

ListTypeRegistrationsResponse'

Fields

registrationTokenList :: Maybe [Text]

A list of extension registration tokens.

Use DescribeTypeRegistration to return detailed information about a type registration request.

nextToken :: Maybe Text

If the request doesn't return all of the remaining results, NextToken is set to a token. To retrieve the next set of results, call this action again and assign that token to the request object's NextToken parameter. If the request returns all results, NextToken is set to null.

httpStatus :: Int

The response's http status code.

newListTypeRegistrationsResponse

Arguments

:: Int
$sel:httpStatus:ListTypeRegistrationsResponse'

-> ListTypeRegistrationsResponse

Create a value of ListTypeRegistrationsResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:registrationTokenList:ListTypeRegistrationsResponse', listTypeRegistrationsResponse_registrationTokenList - A list of extension registration tokens.

Use DescribeTypeRegistration to return detailed information about a type registration request.

$sel:nextToken:ListTypeRegistrations', listTypeRegistrationsResponse_nextToken - If the request doesn't return all of the remaining results, NextToken is set to a token. To retrieve the next set of results, call this action again and assign that token to the request object's NextToken parameter. If the request returns all results, NextToken is set to null.

$sel:httpStatus:ListTypeRegistrationsResponse', listTypeRegistrationsResponse_httpStatus - The response's http status code.

Response Lenses

listTypeRegistrationsResponse_registrationTokenList :: Lens' ListTypeRegistrationsResponse (Maybe [Text])

A list of extension registration tokens.

Use DescribeTypeRegistration to return detailed information about a type registration request.

listTypeRegistrationsResponse_nextToken :: Lens' ListTypeRegistrationsResponse (Maybe Text)

If the request doesn't return all of the remaining results, NextToken is set to a token. To retrieve the next set of results, call this action again and assign that token to the request object's NextToken parameter. If the request returns all results, NextToken is set to null.

listTypeRegistrationsResponse_httpStatus :: Lens' ListTypeRegistrationsResponse Int

The response's http status code.