Amazonka.CloudFormation.SetStackPolicy

Description

Sets a stack policy for a specified stack.

Creating a Request

data SetStackPolicy

The input for the SetStackPolicy action.

See: newSetStackPolicy smart constructor.

Constructors

SetStackPolicy'

Fields

stackPolicyBody :: Maybe Text

Structure containing the stack policy body. For more information, go to Prevent Updates to Stack Resources in the CloudFormation User Guide. You can specify either the StackPolicyBody or the StackPolicyURL parameter, but not both.

stackPolicyURL :: Maybe Text

Location of a file containing the stack policy. The URL must point to a policy (maximum size: 16 KB) located in an S3 bucket in the same Region as the stack. You can specify either the StackPolicyBody or the StackPolicyURL parameter, but not both.

stackName :: Text

The name or unique stack ID that you want to associate a policy with.

newSetStackPolicy

Arguments

:: Text
$sel:stackName:SetStackPolicy'

-> SetStackPolicy

Create a value of SetStackPolicy with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:stackPolicyBody:SetStackPolicy', setStackPolicy_stackPolicyBody - Structure containing the stack policy body. For more information, go to Prevent Updates to Stack Resources in the CloudFormation User Guide. You can specify either the StackPolicyBody or the StackPolicyURL parameter, but not both.

$sel:stackPolicyURL:SetStackPolicy', setStackPolicy_stackPolicyURL - Location of a file containing the stack policy. The URL must point to a policy (maximum size: 16 KB) located in an S3 bucket in the same Region as the stack. You can specify either the StackPolicyBody or the StackPolicyURL parameter, but not both.

$sel:stackName:SetStackPolicy', setStackPolicy_stackName - The name or unique stack ID that you want to associate a policy with.

Request Lenses

setStackPolicy_stackPolicyBody :: Lens' SetStackPolicy (Maybe Text)

Structure containing the stack policy body. For more information, go to Prevent Updates to Stack Resources in the CloudFormation User Guide. You can specify either the StackPolicyBody or the StackPolicyURL parameter, but not both.

setStackPolicy_stackPolicyURL :: Lens' SetStackPolicy (Maybe Text)

Location of a file containing the stack policy. The URL must point to a policy (maximum size: 16 KB) located in an S3 bucket in the same Region as the stack. You can specify either the StackPolicyBody or the StackPolicyURL parameter, but not both.

setStackPolicy_stackName :: Lens' SetStackPolicy Text

The name or unique stack ID that you want to associate a policy with.

Destructuring the Response

data SetStackPolicyResponse

See: newSetStackPolicyResponse smart constructor.

Constructors

SetStackPolicyResponse'

newSetStackPolicyResponse :: SetStackPolicyResponse

Create a value of SetStackPolicyResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.