Amazonka.CloudFormation.Types.DeploymentTargets

Description

Documentation

data DeploymentTargets

Service-managed permissions

The Organizations accounts to which StackSets deploys. StackSets does not deploy stack instances to the organization management account, even if the organization management account is in your organization or in an OU in your organization.

For update operations, you can specify either Accounts or OrganizationalUnitIds. For create and delete operations, specify OrganizationalUnitIds.

See: newDeploymentTargets smart constructor.

Constructors

DeploymentTargets'

Fields

accounts :: Maybe [Text]

The names of one or more Amazon Web Services accounts for which you want to deploy stack set updates.

organizationalUnitIds :: Maybe [Text]

The organization root ID or organizational unit (OU) IDs to which StackSets deploys.

accountsUrl :: Maybe Text

Returns the value of the AccountsUrl property.

newDeploymentTargets :: DeploymentTargets

Create a value of DeploymentTargets with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:accounts:DeploymentTargets', deploymentTargets_accounts - The names of one or more Amazon Web Services accounts for which you want to deploy stack set updates.

$sel:organizationalUnitIds:DeploymentTargets', deploymentTargets_organizationalUnitIds - The organization root ID or organizational unit (OU) IDs to which StackSets deploys.

$sel:accountsUrl:DeploymentTargets', deploymentTargets_accountsUrl - Returns the value of the AccountsUrl property.

deploymentTargets_accounts :: Lens' DeploymentTargets (Maybe [Text])

The names of one or more Amazon Web Services accounts for which you want to deploy stack set updates.

deploymentTargets_organizationalUnitIds :: Lens' DeploymentTargets (Maybe [Text])

The organization root ID or organizational unit (OU) IDs to which StackSets deploys.

deploymentTargets_accountsUrl :: Lens' DeploymentTargets (Maybe Text)

Returns the value of the AccountsUrl property.