Amazonka.CloudFormation.Types.Stack

Description

Documentation

data Stack

The Stack data type.

See: newStack smart constructor.

Constructors

Stack'

Fields

disableRollback :: Maybe Bool

Boolean to enable or disable rollback on stack creation failures:

true: disable rollback

false: enable rollback

lastUpdatedTime :: Maybe ISO8601

The time the stack was last updated. This field will only be returned if the stack has been updated at least once.

rootId :: Maybe Text

For nested stacks--stacks created as resources for another stack--the stack ID of the top-level stack to which the nested stack ultimately belongs.

For more information, see Working with Nested Stacks in the CloudFormation User Guide.

notificationARNs :: Maybe [Text]

SNS topic ARNs to which stack related events are published.

stackStatusReason :: Maybe Text

Success/failure message associated with the stack status.

enableTerminationProtection :: Maybe Bool

Whether termination protection is enabled for the stack.

For nested stacks, termination protection is set on the root stack and cannot be changed directly on the nested stack. For more information, see Protecting a Stack From Being Deleted in the CloudFormation User Guide.

driftInformation :: Maybe StackDriftInformation

Information on whether a stack's actual configuration differs, or has drifted, from it's expected configuration, as defined in the stack template and any values specified as template parameters. For more information, see Detecting Unregulated Configuration Changes to Stacks and Resources.

changeSetId :: Maybe Text

The unique ID of the change set.

deletionTime :: Maybe ISO8601

The time the stack was deleted.

outputs :: Maybe [Output]

A list of output structures.

parameters :: Maybe [Parameter]

A list of Parameter structures.

stackId :: Maybe Text

Unique identifier of the stack.

description :: Maybe Text

A user-defined description associated with the stack.

capabilities :: Maybe [Capability]

The capabilities allowed in the stack.

rollbackConfiguration :: Maybe RollbackConfiguration

The rollback triggers for CloudFormation to monitor during stack creation and updating operations, and for the specified monitoring period afterwards.

tags :: Maybe [Tag]

A list of Tags that specify information about the stack.

timeoutInMinutes :: Maybe Natural

The amount of time within which stack creation should complete.

parentId :: Maybe Text

For nested stacks--stacks created as resources for another stack--the stack ID of the direct parent of this stack. For the first level of nested stacks, the root stack is also the parent stack.

For more information, see Working with Nested Stacks in the CloudFormation User Guide.

roleARN :: Maybe Text

The Amazon Resource Name (ARN) of an Identity and Access Management (IAM) role that is associated with the stack. During a stack operation, CloudFormation uses this role's credentials to make calls on your behalf.

stackName :: Text

The name associated with the stack.

creationTime :: ISO8601

The time at which the stack was created.

stackStatus :: StackStatus

Current status of the stack.

newStack

Arguments

:: Text
$sel:stackName:Stack'

-> UTCTime
$sel:creationTime:Stack'

-> StackStatus
$sel:stackStatus:Stack'

-> Stack

Create a value of Stack with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:disableRollback:Stack', stack_disableRollback - Boolean to enable or disable rollback on stack creation failures:

true: disable rollback

false: enable rollback

$sel:lastUpdatedTime:Stack', stack_lastUpdatedTime - The time the stack was last updated. This field will only be returned if the stack has been updated at least once.

$sel:rootId:Stack', stack_rootId - For nested stacks--stacks created as resources for another stack--the stack ID of the top-level stack to which the nested stack ultimately belongs.

For more information, see Working with Nested Stacks in the CloudFormation User Guide.

$sel:notificationARNs:Stack', stack_notificationARNs - SNS topic ARNs to which stack related events are published.

$sel:stackStatusReason:Stack', stack_stackStatusReason - Success/failure message associated with the stack status.

$sel:enableTerminationProtection:Stack', stack_enableTerminationProtection - Whether termination protection is enabled for the stack.

For nested stacks, termination protection is set on the root stack and cannot be changed directly on the nested stack. For more information, see Protecting a Stack From Being Deleted in the CloudFormation User Guide.

$sel:driftInformation:Stack', stack_driftInformation - Information on whether a stack's actual configuration differs, or has drifted, from it's expected configuration, as defined in the stack template and any values specified as template parameters. For more information, see Detecting Unregulated Configuration Changes to Stacks and Resources.

$sel:changeSetId:Stack', stack_changeSetId - The unique ID of the change set.

$sel:deletionTime:Stack', stack_deletionTime - The time the stack was deleted.

$sel:outputs:Stack', stack_outputs - A list of output structures.

$sel:parameters:Stack', stack_parameters - A list of Parameter structures.

$sel:stackId:Stack', stack_stackId - Unique identifier of the stack.

$sel:description:Stack', stack_description - A user-defined description associated with the stack.

$sel:capabilities:Stack', stack_capabilities - The capabilities allowed in the stack.

$sel:rollbackConfiguration:Stack', stack_rollbackConfiguration - The rollback triggers for CloudFormation to monitor during stack creation and updating operations, and for the specified monitoring period afterwards.

$sel:tags:Stack', stack_tags - A list of Tags that specify information about the stack.

$sel:timeoutInMinutes:Stack', stack_timeoutInMinutes - The amount of time within which stack creation should complete.

$sel:parentId:Stack', stack_parentId - For nested stacks--stacks created as resources for another stack--the stack ID of the direct parent of this stack. For the first level of nested stacks, the root stack is also the parent stack.

For more information, see Working with Nested Stacks in the CloudFormation User Guide.

$sel:roleARN:Stack', stack_roleARN - The Amazon Resource Name (ARN) of an Identity and Access Management (IAM) role that is associated with the stack. During a stack operation, CloudFormation uses this role's credentials to make calls on your behalf.

$sel:stackName:Stack', stack_stackName - The name associated with the stack.

$sel:creationTime:Stack', stack_creationTime - The time at which the stack was created.

$sel:stackStatus:Stack', stack_stackStatus - Current status of the stack.

stack_disableRollback :: Lens' Stack (Maybe Bool)

Boolean to enable or disable rollback on stack creation failures:

true: disable rollback

false: enable rollback

stack_lastUpdatedTime :: Lens' Stack (Maybe UTCTime)

The time the stack was last updated. This field will only be returned if the stack has been updated at least once.

stack_rootId :: Lens' Stack (Maybe Text)

For nested stacks--stacks created as resources for another stack--the stack ID of the top-level stack to which the nested stack ultimately belongs.

For more information, see Working with Nested Stacks in the CloudFormation User Guide.

stack_notificationARNs :: Lens' Stack (Maybe [Text])

SNS topic ARNs to which stack related events are published.

stack_stackStatusReason :: Lens' Stack (Maybe Text)

Success/failure message associated with the stack status.

stack_enableTerminationProtection :: Lens' Stack (Maybe Bool)

Whether termination protection is enabled for the stack.

For nested stacks, termination protection is set on the root stack and cannot be changed directly on the nested stack. For more information, see Protecting a Stack From Being Deleted in the CloudFormation User Guide.

stack_driftInformation :: Lens' Stack (Maybe StackDriftInformation)

Information on whether a stack's actual configuration differs, or has drifted, from it's expected configuration, as defined in the stack template and any values specified as template parameters. For more information, see Detecting Unregulated Configuration Changes to Stacks and Resources.

stack_changeSetId :: Lens' Stack (Maybe Text)

The unique ID of the change set.

stack_deletionTime :: Lens' Stack (Maybe UTCTime)

The time the stack was deleted.

stack_outputs :: Lens' Stack (Maybe [Output])

A list of output structures.

stack_parameters :: Lens' Stack (Maybe [Parameter])

A list of Parameter structures.

stack_stackId :: Lens' Stack (Maybe Text)

Unique identifier of the stack.

stack_description :: Lens' Stack (Maybe Text)

A user-defined description associated with the stack.

stack_capabilities :: Lens' Stack (Maybe [Capability])

The capabilities allowed in the stack.

stack_rollbackConfiguration :: Lens' Stack (Maybe RollbackConfiguration)

The rollback triggers for CloudFormation to monitor during stack creation and updating operations, and for the specified monitoring period afterwards.

stack_tags :: Lens' Stack (Maybe [Tag])

A list of Tags that specify information about the stack.

stack_timeoutInMinutes :: Lens' Stack (Maybe Natural)

The amount of time within which stack creation should complete.

stack_parentId :: Lens' Stack (Maybe Text)

For nested stacks--stacks created as resources for another stack--the stack ID of the direct parent of this stack. For the first level of nested stacks, the root stack is also the parent stack.

For more information, see Working with Nested Stacks in the CloudFormation User Guide.

stack_roleARN :: Lens' Stack (Maybe Text)

The Amazon Resource Name (ARN) of an Identity and Access Management (IAM) role that is associated with the stack. During a stack operation, CloudFormation uses this role's credentials to make calls on your behalf.

stack_stackName :: Lens' Stack Text

The name associated with the stack.

stack_creationTime :: Lens' Stack UTCTime

The time at which the stack was created.

stack_stackStatus :: Lens' Stack StackStatus

Current status of the stack.