Amazonka.CloudFormation.Types.StackDriftInformationSummary

Description

Documentation

data StackDriftInformationSummary

Contains information about whether the stack's actual configuration differs, or has drifted, from its expected configuration, as defined in the stack template and any values specified as template parameters. A stack is considered to have drifted if one or more of its resources have drifted.

See: newStackDriftInformationSummary smart constructor.

Constructors

StackDriftInformationSummary'

Fields

lastCheckTimestamp :: Maybe ISO8601

Most recent time when a drift detection operation was initiated on the stack, or any of its individual resources that support drift detection.

stackDriftStatus :: StackDriftStatus

Status of the stack's actual configuration compared to its expected template configuration.

DRIFTED: The stack differs from its expected template configuration. A stack is considered to have drifted if one or more of its resources have drifted.

NOT_CHECKED: CloudFormation has not checked if the stack differs from its expected template configuration.

IN_SYNC: The stack's actual configuration matches its expected template configuration.

UNKNOWN: This value is reserved for future use.

newStackDriftInformationSummary

Arguments

:: StackDriftStatus
$sel:stackDriftStatus:StackDriftInformationSummary'

-> StackDriftInformationSummary

Create a value of StackDriftInformationSummary with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:lastCheckTimestamp:StackDriftInformationSummary', stackDriftInformationSummary_lastCheckTimestamp - Most recent time when a drift detection operation was initiated on the stack, or any of its individual resources that support drift detection.

$sel:stackDriftStatus:StackDriftInformationSummary', stackDriftInformationSummary_stackDriftStatus - Status of the stack's actual configuration compared to its expected template configuration.

DRIFTED: The stack differs from its expected template configuration. A stack is considered to have drifted if one or more of its resources have drifted.

NOT_CHECKED: CloudFormation has not checked if the stack differs from its expected template configuration.

IN_SYNC: The stack's actual configuration matches its expected template configuration.

UNKNOWN: This value is reserved for future use.

stackDriftInformationSummary_lastCheckTimestamp :: Lens' StackDriftInformationSummary (Maybe UTCTime)

Most recent time when a drift detection operation was initiated on the stack, or any of its individual resources that support drift detection.

stackDriftInformationSummary_stackDriftStatus :: Lens' StackDriftInformationSummary StackDriftStatus

Status of the stack's actual configuration compared to its expected template configuration.

DRIFTED: The stack differs from its expected template configuration. A stack is considered to have drifted if one or more of its resources have drifted.

NOT_CHECKED: CloudFormation has not checked if the stack differs from its expected template configuration.

IN_SYNC: The stack's actual configuration matches its expected template configuration.

UNKNOWN: This value is reserved for future use.