Amazonka.CloudFormation.Types.StackInstance

Description

Documentation

data StackInstance

An CloudFormation stack, in a specific account and Region, that's part of a stack set operation. A stack instance is a reference to an attempted or actual stack in a given account within a given Region. A stack instance can exist without a stack—for example, if the stack couldn't be created for some reason. A stack instance is associated with only one stack set. Each stack instance contains the ID of its associated stack set, as well as the ID of the actual stack and the stack status.

See: newStackInstance smart constructor.

Constructors

StackInstance'

Fields

status :: Maybe StackInstanceStatus

The status of the stack instance, in terms of its synchronization with its associated stack set.

INOPERABLE: A DeleteStackInstances operation has failed and left the stack in an unstable state. Stacks in this state are excluded from further UpdateStackSet operations. You might need to perform a DeleteStackInstances operation, with RetainStacks set to true, to delete the stack instance, and then delete the stack manually.

OUTDATED: The stack isn't currently up to date with the stack set because:

The associated stack failed during a CreateStackSet or UpdateStackSet operation.

The stack was part of a CreateStackSet or UpdateStackSet operation that failed or was stopped before the stack was created or updated.

CURRENT: The stack is currently up to date with the stack set.

lastDriftCheckTimestamp :: Maybe ISO8601

Most recent time when CloudFormation performed a drift detection operation on the stack instance. This value will be NULL for any stack instance on which drift detection has not yet been performed.

account :: Maybe Text

Self-managed permissions

The name of the Amazon Web Services account that the stack instance is associated with.

driftStatus :: Maybe StackDriftStatus

Status of the stack instance's actual configuration compared to the expected template and parameter configuration of the stack set to which it belongs.

DRIFTED: The stack differs from the expected template and parameter configuration of the stack set to which it belongs. A stack instance is considered to have drifted if one or more of the resources in the associated stack have drifted.

NOT_CHECKED: CloudFormation has not checked if the stack instance differs from its expected stack set configuration.

IN_SYNC: The stack instance's actual configuration matches its expected stack set configuration.

UNKNOWN: This value is reserved for future use.

organizationalUnitId :: Maybe Text

Service-managed permissions

The organization root ID or organizational unit (OU) IDs that you specified for DeploymentTargets.

region :: Maybe Text

The name of the Amazon Web Services Region that the stack instance is associated with.

statusReason :: Maybe Text

The explanation for the specific status code that is assigned to this stack instance.

stackId :: Maybe Text

The ID of the stack instance.

stackInstanceStatus :: Maybe StackInstanceComprehensiveStatus

The detailed status of the stack instance.

parameterOverrides :: Maybe [Parameter]

A list of parameters from the stack set template whose values have been overridden in this stack instance.

stackSetId :: Maybe Text

The name or unique ID of the stack set that the stack instance is associated with.

newStackInstance :: StackInstance

Create a value of StackInstance with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:status:StackInstance', stackInstance_status - The status of the stack instance, in terms of its synchronization with its associated stack set.

INOPERABLE: A DeleteStackInstances operation has failed and left the stack in an unstable state. Stacks in this state are excluded from further UpdateStackSet operations. You might need to perform a DeleteStackInstances operation, with RetainStacks set to true, to delete the stack instance, and then delete the stack manually.

OUTDATED: The stack isn't currently up to date with the stack set because:

The associated stack failed during a CreateStackSet or UpdateStackSet operation.

The stack was part of a CreateStackSet or UpdateStackSet operation that failed or was stopped before the stack was created or updated.

CURRENT: The stack is currently up to date with the stack set.

$sel:lastDriftCheckTimestamp:StackInstance', stackInstance_lastDriftCheckTimestamp - Most recent time when CloudFormation performed a drift detection operation on the stack instance. This value will be NULL for any stack instance on which drift detection has not yet been performed.

$sel:account:StackInstance', stackInstance_account - [Self-managed permissions] The name of the Amazon Web Services account that the stack instance is associated with.

$sel:driftStatus:StackInstance', stackInstance_driftStatus - Status of the stack instance's actual configuration compared to the expected template and parameter configuration of the stack set to which it belongs.

DRIFTED: The stack differs from the expected template and parameter configuration of the stack set to which it belongs. A stack instance is considered to have drifted if one or more of the resources in the associated stack have drifted.

NOT_CHECKED: CloudFormation has not checked if the stack instance differs from its expected stack set configuration.

IN_SYNC: The stack instance's actual configuration matches its expected stack set configuration.

UNKNOWN: This value is reserved for future use.

$sel:organizationalUnitId:StackInstance', stackInstance_organizationalUnitId - [Service-managed permissions] The organization root ID or organizational unit (OU) IDs that you specified for DeploymentTargets.

$sel:region:StackInstance', stackInstance_region - The name of the Amazon Web Services Region that the stack instance is associated with.

$sel:statusReason:StackInstance', stackInstance_statusReason - The explanation for the specific status code that is assigned to this stack instance.

$sel:stackId:StackInstance', stackInstance_stackId - The ID of the stack instance.

$sel:stackInstanceStatus:StackInstance', stackInstance_stackInstanceStatus - The detailed status of the stack instance.

$sel:parameterOverrides:StackInstance', stackInstance_parameterOverrides - A list of parameters from the stack set template whose values have been overridden in this stack instance.

$sel:stackSetId:StackInstance', stackInstance_stackSetId - The name or unique ID of the stack set that the stack instance is associated with.

stackInstance_status :: Lens' StackInstance (Maybe StackInstanceStatus)

The status of the stack instance, in terms of its synchronization with its associated stack set.

INOPERABLE: A DeleteStackInstances operation has failed and left the stack in an unstable state. Stacks in this state are excluded from further UpdateStackSet operations. You might need to perform a DeleteStackInstances operation, with RetainStacks set to true, to delete the stack instance, and then delete the stack manually.

OUTDATED: The stack isn't currently up to date with the stack set because:

The associated stack failed during a CreateStackSet or UpdateStackSet operation.

The stack was part of a CreateStackSet or UpdateStackSet operation that failed or was stopped before the stack was created or updated.

CURRENT: The stack is currently up to date with the stack set.

stackInstance_lastDriftCheckTimestamp :: Lens' StackInstance (Maybe UTCTime)

Most recent time when CloudFormation performed a drift detection operation on the stack instance. This value will be NULL for any stack instance on which drift detection has not yet been performed.

stackInstance_account :: Lens' StackInstance (Maybe Text)

Self-managed permissions

The name of the Amazon Web Services account that the stack instance is associated with.

stackInstance_driftStatus :: Lens' StackInstance (Maybe StackDriftStatus)

Status of the stack instance's actual configuration compared to the expected template and parameter configuration of the stack set to which it belongs.

DRIFTED: The stack differs from the expected template and parameter configuration of the stack set to which it belongs. A stack instance is considered to have drifted if one or more of the resources in the associated stack have drifted.

NOT_CHECKED: CloudFormation has not checked if the stack instance differs from its expected stack set configuration.

IN_SYNC: The stack instance's actual configuration matches its expected stack set configuration.

UNKNOWN: This value is reserved for future use.

stackInstance_organizationalUnitId :: Lens' StackInstance (Maybe Text)

Service-managed permissions

The organization root ID or organizational unit (OU) IDs that you specified for DeploymentTargets.

stackInstance_region :: Lens' StackInstance (Maybe Text)

The name of the Amazon Web Services Region that the stack instance is associated with.

stackInstance_statusReason :: Lens' StackInstance (Maybe Text)

The explanation for the specific status code that is assigned to this stack instance.

stackInstance_stackId :: Lens' StackInstance (Maybe Text)

The ID of the stack instance.

stackInstance_stackInstanceStatus :: Lens' StackInstance (Maybe StackInstanceComprehensiveStatus)

The detailed status of the stack instance.

stackInstance_parameterOverrides :: Lens' StackInstance (Maybe [Parameter])

A list of parameters from the stack set template whose values have been overridden in this stack instance.

stackInstance_stackSetId :: Lens' StackInstance (Maybe Text)

The name or unique ID of the stack set that the stack instance is associated with.