Amazonka.CloudFormation.Types.StackResourceSummary

Description

Documentation

data StackResourceSummary

Contains high-level information about the specified stack resource.

See: newStackResourceSummary smart constructor.

Constructors

StackResourceSummary'

Fields

physicalResourceId :: Maybe Text

The name or unique identifier that corresponds to a physical instance ID of the resource.

resourceStatusReason :: Maybe Text

Success/failure message associated with the resource.

driftInformation :: Maybe StackResourceDriftInformationSummary

Information about whether the resource's actual configuration differs, or has drifted, from its expected configuration, as defined in the stack template and any values specified as template parameters. For more information, see Detecting Unregulated Configuration Changes to Stacks and Resources.

moduleInfo :: Maybe ModuleInfo

Contains information about the module from which the resource was created, if the resource was created from a module included in the stack template.

logicalResourceId :: Text

The logical name of the resource specified in the template.

resourceType :: Text

Type of resource. (For more information, go to Amazon Web Services Resource Types Reference in the CloudFormation User Guide.)

lastUpdatedTimestamp :: ISO8601

Time the status was updated.

resourceStatus :: ResourceStatus

Current status of the resource.

newStackResourceSummary

Arguments

:: Text
$sel:logicalResourceId:StackResourceSummary'

-> Text
$sel:resourceType:StackResourceSummary'

-> UTCTime
$sel:lastUpdatedTimestamp:StackResourceSummary'

-> ResourceStatus
$sel:resourceStatus:StackResourceSummary'

-> StackResourceSummary

Create a value of StackResourceSummary with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:physicalResourceId:StackResourceSummary', stackResourceSummary_physicalResourceId - The name or unique identifier that corresponds to a physical instance ID of the resource.

$sel:resourceStatusReason:StackResourceSummary', stackResourceSummary_resourceStatusReason - Success/failure message associated with the resource.

$sel:driftInformation:StackResourceSummary', stackResourceSummary_driftInformation - Information about whether the resource's actual configuration differs, or has drifted, from its expected configuration, as defined in the stack template and any values specified as template parameters. For more information, see Detecting Unregulated Configuration Changes to Stacks and Resources.

$sel:moduleInfo:StackResourceSummary', stackResourceSummary_moduleInfo - Contains information about the module from which the resource was created, if the resource was created from a module included in the stack template.

$sel:logicalResourceId:StackResourceSummary', stackResourceSummary_logicalResourceId - The logical name of the resource specified in the template.

$sel:resourceType:StackResourceSummary', stackResourceSummary_resourceType - Type of resource. (For more information, go to Amazon Web Services Resource Types Reference in the CloudFormation User Guide.)

$sel:lastUpdatedTimestamp:StackResourceSummary', stackResourceSummary_lastUpdatedTimestamp - Time the status was updated.

$sel:resourceStatus:StackResourceSummary', stackResourceSummary_resourceStatus - Current status of the resource.

stackResourceSummary_physicalResourceId :: Lens' StackResourceSummary (Maybe Text)

The name or unique identifier that corresponds to a physical instance ID of the resource.

stackResourceSummary_resourceStatusReason :: Lens' StackResourceSummary (Maybe Text)

Success/failure message associated with the resource.

stackResourceSummary_driftInformation :: Lens' StackResourceSummary (Maybe StackResourceDriftInformationSummary)

Information about whether the resource's actual configuration differs, or has drifted, from its expected configuration, as defined in the stack template and any values specified as template parameters. For more information, see Detecting Unregulated Configuration Changes to Stacks and Resources.

stackResourceSummary_moduleInfo :: Lens' StackResourceSummary (Maybe ModuleInfo)

Contains information about the module from which the resource was created, if the resource was created from a module included in the stack template.

stackResourceSummary_logicalResourceId :: Lens' StackResourceSummary Text

The logical name of the resource specified in the template.

stackResourceSummary_resourceType :: Lens' StackResourceSummary Text

Type of resource. (For more information, go to Amazon Web Services Resource Types Reference in the CloudFormation User Guide.)

stackResourceSummary_lastUpdatedTimestamp :: Lens' StackResourceSummary UTCTime

Time the status was updated.

stackResourceSummary_resourceStatus :: Lens' StackResourceSummary ResourceStatus

Current status of the resource.