Amazonka.CloudFormation.Types.StackSet

Description

Documentation

data StackSet

A structure that contains information about a stack set. A stack set enables you to provision stacks into Amazon Web Services accounts and across Regions by using a single CloudFormation template. In the stack set, you specify the template to use, as well as any parameters and capabilities that the template requires.

See: newStackSet smart constructor.

Constructors

StackSet'

Fields

stackSetDriftDetectionDetails :: Maybe StackSetDriftDetectionDetails

Detailed information about the drift status of the stack set.

For stack sets, contains information about the last completed drift operation performed on the stack set. Information about drift operations currently in progress is not included.

status :: Maybe StackSetStatus

The status of the stack set.

administrationRoleARN :: Maybe Text

The Amazon Resource Number (ARN) of the IAM role used to create or update the stack set.

Use customized administrator roles to control which users or groups can manage specific stack sets within the same administrator account. For more information, see Prerequisites: Granting Permissions for Stack Set Operations in the CloudFormation User Guide.

autoDeployment :: Maybe AutoDeployment

Service-managed permissions

Describes whether StackSets automatically deploys to Organizations accounts that are added to a target organization or organizational unit (OU).

organizationalUnitIds :: Maybe [Text]

Service-managed permissions

The organization root ID or organizational unit (OU) IDs that you specified for DeploymentTargets.

stackSetARN :: Maybe Text

The Amazon Resource Number (ARN) of the stack set.

permissionModel :: Maybe PermissionModels

Describes how the IAM roles required for stack set operations are created.

With self-managed permissions, you must create the administrator and execution roles required to deploy to target accounts. For more information, see Grant Self-Managed Stack Set Permissions.

With service-managed permissions, StackSets automatically creates the IAM roles required to deploy to accounts managed by Organizations. For more information, see Grant Service-Managed Stack Set Permissions.

parameters :: Maybe [Parameter]

A list of input parameters for a stack set.

templateBody :: Maybe Text

The structure that contains the body of the template that was used to create or update the stack set.

stackSetName :: Maybe Text

The name that's associated with the stack set.

description :: Maybe Text

A description of the stack set that you specify when the stack set is created or updated.

capabilities :: Maybe [Capability]

The capabilities that are allowed in the stack set. Some stack set templates might include resources that can affect permissions in your Amazon Web Services account—for example, by creating new Identity and Access Management (IAM) users. For more information, see Acknowledging IAM Resources in CloudFormation Templates.

tags :: Maybe [Tag]

A list of tags that specify information about the stack set. A maximum number of 50 tags can be specified.

stackSetId :: Maybe Text

The ID of the stack set.

executionRoleName :: Maybe Text

The name of the IAM execution role used to create or update the stack set.

Use customized execution roles to control which stack resources users and groups can include in their stack sets.

newStackSet :: StackSet

Create a value of StackSet with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:stackSetDriftDetectionDetails:StackSet', stackSet_stackSetDriftDetectionDetails - Detailed information about the drift status of the stack set.

For stack sets, contains information about the last completed drift operation performed on the stack set. Information about drift operations currently in progress is not included.

$sel:status:StackSet', stackSet_status - The status of the stack set.

$sel:administrationRoleARN:StackSet', stackSet_administrationRoleARN - The Amazon Resource Number (ARN) of the IAM role used to create or update the stack set.

Use customized administrator roles to control which users or groups can manage specific stack sets within the same administrator account. For more information, see Prerequisites: Granting Permissions for Stack Set Operations in the CloudFormation User Guide.

$sel:autoDeployment:StackSet', stackSet_autoDeployment - [Service-managed permissions] Describes whether StackSets automatically deploys to Organizations accounts that are added to a target organization or organizational unit (OU).

$sel:organizationalUnitIds:StackSet', stackSet_organizationalUnitIds - [Service-managed permissions] The organization root ID or organizational unit (OU) IDs that you specified for DeploymentTargets.

$sel:stackSetARN:StackSet', stackSet_stackSetARN - The Amazon Resource Number (ARN) of the stack set.

$sel:permissionModel:StackSet', stackSet_permissionModel - Describes how the IAM roles required for stack set operations are created.

With self-managed permissions, you must create the administrator and execution roles required to deploy to target accounts. For more information, see Grant Self-Managed Stack Set Permissions.

With service-managed permissions, StackSets automatically creates the IAM roles required to deploy to accounts managed by Organizations. For more information, see Grant Service-Managed Stack Set Permissions.

$sel:parameters:StackSet', stackSet_parameters - A list of input parameters for a stack set.

$sel:templateBody:StackSet', stackSet_templateBody - The structure that contains the body of the template that was used to create or update the stack set.

$sel:stackSetName:StackSet', stackSet_stackSetName - The name that's associated with the stack set.

$sel:description:StackSet', stackSet_description - A description of the stack set that you specify when the stack set is created or updated.

$sel:capabilities:StackSet', stackSet_capabilities - The capabilities that are allowed in the stack set. Some stack set templates might include resources that can affect permissions in your Amazon Web Services account—for example, by creating new Identity and Access Management (IAM) users. For more information, see Acknowledging IAM Resources in CloudFormation Templates.

$sel:tags:StackSet', stackSet_tags - A list of tags that specify information about the stack set. A maximum number of 50 tags can be specified.

$sel:stackSetId:StackSet', stackSet_stackSetId - The ID of the stack set.

$sel:executionRoleName:StackSet', stackSet_executionRoleName - The name of the IAM execution role used to create or update the stack set.

Use customized execution roles to control which stack resources users and groups can include in their stack sets.

stackSet_stackSetDriftDetectionDetails :: Lens' StackSet (Maybe StackSetDriftDetectionDetails)

Detailed information about the drift status of the stack set.

For stack sets, contains information about the last completed drift operation performed on the stack set. Information about drift operations currently in progress is not included.

stackSet_status :: Lens' StackSet (Maybe StackSetStatus)

The status of the stack set.

stackSet_administrationRoleARN :: Lens' StackSet (Maybe Text)

The Amazon Resource Number (ARN) of the IAM role used to create or update the stack set.

Use customized administrator roles to control which users or groups can manage specific stack sets within the same administrator account. For more information, see Prerequisites: Granting Permissions for Stack Set Operations in the CloudFormation User Guide.

stackSet_autoDeployment :: Lens' StackSet (Maybe AutoDeployment)

Service-managed permissions

Describes whether StackSets automatically deploys to Organizations accounts that are added to a target organization or organizational unit (OU).

stackSet_organizationalUnitIds :: Lens' StackSet (Maybe [Text])

Service-managed permissions

The organization root ID or organizational unit (OU) IDs that you specified for DeploymentTargets.

stackSet_stackSetARN :: Lens' StackSet (Maybe Text)

The Amazon Resource Number (ARN) of the stack set.

stackSet_permissionModel :: Lens' StackSet (Maybe PermissionModels)

Describes how the IAM roles required for stack set operations are created.

With self-managed permissions, you must create the administrator and execution roles required to deploy to target accounts. For more information, see Grant Self-Managed Stack Set Permissions.

With service-managed permissions, StackSets automatically creates the IAM roles required to deploy to accounts managed by Organizations. For more information, see Grant Service-Managed Stack Set Permissions.

stackSet_parameters :: Lens' StackSet (Maybe [Parameter])

A list of input parameters for a stack set.

stackSet_templateBody :: Lens' StackSet (Maybe Text)

The structure that contains the body of the template that was used to create or update the stack set.

stackSet_stackSetName :: Lens' StackSet (Maybe Text)

The name that's associated with the stack set.

stackSet_description :: Lens' StackSet (Maybe Text)

A description of the stack set that you specify when the stack set is created or updated.

stackSet_capabilities :: Lens' StackSet (Maybe [Capability])

The capabilities that are allowed in the stack set. Some stack set templates might include resources that can affect permissions in your Amazon Web Services account—for example, by creating new Identity and Access Management (IAM) users. For more information, see Acknowledging IAM Resources in CloudFormation Templates.

stackSet_tags :: Lens' StackSet (Maybe [Tag])

A list of tags that specify information about the stack set. A maximum number of 50 tags can be specified.

stackSet_stackSetId :: Lens' StackSet (Maybe Text)

The ID of the stack set.

stackSet_executionRoleName :: Lens' StackSet (Maybe Text)

The name of the IAM execution role used to create or update the stack set.

Use customized execution roles to control which stack resources users and groups can include in their stack sets.