Amazonka.CloudFormation.Types.StackSetOperation

Description

Documentation

data StackSetOperation

The structure that contains information about a stack set operation.

See: newStackSetOperation smart constructor.

Constructors

StackSetOperation'

Fields

stackSetDriftDetectionDetails :: Maybe StackSetDriftDetectionDetails

Detailed information about the drift status of the stack set. This includes information about drift operations currently being performed on the stack set.

this information will only be present for stack set operations whose Action type is DETECT_DRIFT.

For more information, see Detecting Unmanaged Changes in Stack Sets in the CloudFormation User Guide.

status :: Maybe StackSetOperationStatus

The status of the operation.

FAILED: The operation exceeded the specified failure tolerance. The failure tolerance value that you've set for an operation is applied for each Region during stack create and update operations. If the number of failed stacks within a Region exceeds the failure tolerance, the status of the operation in the Region is set to FAILED. This in turn sets the status of the operation as a whole to FAILED, and CloudFormation cancels the operation in any remaining Regions.

QUEUED: [Service-managed permissions] For automatic deployments that require a sequence of operations, the operation is queued to be performed. For more information, see the stack set operation status codes in the CloudFormation User Guide.

RUNNING: The operation is currently being performed.

STOPPED: The user has cancelled the operation.

STOPPING: The operation is in the process of stopping, at user request.

SUCCEEDED: The operation completed creating or updating all the specified stacks without exceeding the failure tolerance for the operation.

administrationRoleARN :: Maybe Text

The Amazon Resource Number (ARN) of the IAM role used to perform this stack set operation.

Use customized administrator roles to control which users or groups can manage specific stack sets within the same administrator account. For more information, see Define Permissions for Multiple Administrators in the CloudFormation User Guide.

action :: Maybe StackSetOperationAction

The type of stack set operation: CREATE, UPDATE, or DELETE. Create and delete operations affect only the specified stack set instances that are associated with the specified stack set. Update operations affect both the stack set itself, as well as all associated stack set instances.

endTimestamp :: Maybe ISO8601

The time at which the stack set operation ended, across all accounts and Regions specified. Note that this doesn't necessarily mean that the stack set operation was successful, or even attempted, in each account or Region.

creationTimestamp :: Maybe ISO8601

The time at which the operation was initiated. Note that the creation times for the stack set operation might differ from the creation time of the individual stacks themselves. This is because CloudFormation needs to perform preparatory work for the operation, such as dispatching the work to the requested Regions, before actually creating the first stacks.

operationPreferences :: Maybe StackSetOperationPreferences

The preferences for how CloudFormation performs this stack set operation.

operationId :: Maybe Text

The unique ID of a stack set operation.

retainStacks :: Maybe Bool

For stack set operations of action type DELETE, specifies whether to remove the stack instances from the specified stack set, but doesn't delete the stacks. You can't reassociate a retained stack, or add an existing, saved stack to a new stack set.

deploymentTargets :: Maybe DeploymentTargets

Service-managed permissions

The Organizations accounts affected by the stack operation.

stackSetId :: Maybe Text

The ID of the stack set.

executionRoleName :: Maybe Text

The name of the IAM execution role used to create or update the stack set.

Use customized execution roles to control which stack resources users and groups can include in their stack sets.

newStackSetOperation :: StackSetOperation

Create a value of StackSetOperation with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:stackSetDriftDetectionDetails:StackSetOperation', stackSetOperation_stackSetDriftDetectionDetails - Detailed information about the drift status of the stack set. This includes information about drift operations currently being performed on the stack set.

this information will only be present for stack set operations whose Action type is DETECT_DRIFT.

For more information, see Detecting Unmanaged Changes in Stack Sets in the CloudFormation User Guide.

$sel:status:StackSetOperation', stackSetOperation_status - The status of the operation.

FAILED: The operation exceeded the specified failure tolerance. The failure tolerance value that you've set for an operation is applied for each Region during stack create and update operations. If the number of failed stacks within a Region exceeds the failure tolerance, the status of the operation in the Region is set to FAILED. This in turn sets the status of the operation as a whole to FAILED, and CloudFormation cancels the operation in any remaining Regions.

QUEUED: [Service-managed permissions] For automatic deployments that require a sequence of operations, the operation is queued to be performed. For more information, see the stack set operation status codes in the CloudFormation User Guide.

RUNNING: The operation is currently being performed.

STOPPED: The user has cancelled the operation.

STOPPING: The operation is in the process of stopping, at user request.

SUCCEEDED: The operation completed creating or updating all the specified stacks without exceeding the failure tolerance for the operation.

$sel:administrationRoleARN:StackSetOperation', stackSetOperation_administrationRoleARN - The Amazon Resource Number (ARN) of the IAM role used to perform this stack set operation.

Use customized administrator roles to control which users or groups can manage specific stack sets within the same administrator account. For more information, see Define Permissions for Multiple Administrators in the CloudFormation User Guide.

$sel:action:StackSetOperation', stackSetOperation_action - The type of stack set operation: CREATE, UPDATE, or DELETE. Create and delete operations affect only the specified stack set instances that are associated with the specified stack set. Update operations affect both the stack set itself, as well as all associated stack set instances.

$sel:endTimestamp:StackSetOperation', stackSetOperation_endTimestamp - The time at which the stack set operation ended, across all accounts and Regions specified. Note that this doesn't necessarily mean that the stack set operation was successful, or even attempted, in each account or Region.

$sel:creationTimestamp:StackSetOperation', stackSetOperation_creationTimestamp - The time at which the operation was initiated. Note that the creation times for the stack set operation might differ from the creation time of the individual stacks themselves. This is because CloudFormation needs to perform preparatory work for the operation, such as dispatching the work to the requested Regions, before actually creating the first stacks.

$sel:operationPreferences:StackSetOperation', stackSetOperation_operationPreferences - The preferences for how CloudFormation performs this stack set operation.

$sel:operationId:StackSetOperation', stackSetOperation_operationId - The unique ID of a stack set operation.

$sel:retainStacks:StackSetOperation', stackSetOperation_retainStacks - For stack set operations of action type DELETE, specifies whether to remove the stack instances from the specified stack set, but doesn't delete the stacks. You can't reassociate a retained stack, or add an existing, saved stack to a new stack set.

$sel:deploymentTargets:StackSetOperation', stackSetOperation_deploymentTargets - [Service-managed permissions] The Organizations accounts affected by the stack operation.

$sel:stackSetId:StackSetOperation', stackSetOperation_stackSetId - The ID of the stack set.

$sel:executionRoleName:StackSetOperation', stackSetOperation_executionRoleName - The name of the IAM execution role used to create or update the stack set.

Use customized execution roles to control which stack resources users and groups can include in their stack sets.

stackSetOperation_stackSetDriftDetectionDetails :: Lens' StackSetOperation (Maybe StackSetDriftDetectionDetails)

Detailed information about the drift status of the stack set. This includes information about drift operations currently being performed on the stack set.

this information will only be present for stack set operations whose Action type is DETECT_DRIFT.

For more information, see Detecting Unmanaged Changes in Stack Sets in the CloudFormation User Guide.

stackSetOperation_status :: Lens' StackSetOperation (Maybe StackSetOperationStatus)

The status of the operation.

FAILED: The operation exceeded the specified failure tolerance. The failure tolerance value that you've set for an operation is applied for each Region during stack create and update operations. If the number of failed stacks within a Region exceeds the failure tolerance, the status of the operation in the Region is set to FAILED. This in turn sets the status of the operation as a whole to FAILED, and CloudFormation cancels the operation in any remaining Regions.

QUEUED: [Service-managed permissions] For automatic deployments that require a sequence of operations, the operation is queued to be performed. For more information, see the stack set operation status codes in the CloudFormation User Guide.

RUNNING: The operation is currently being performed.

STOPPED: The user has cancelled the operation.

STOPPING: The operation is in the process of stopping, at user request.

SUCCEEDED: The operation completed creating or updating all the specified stacks without exceeding the failure tolerance for the operation.

stackSetOperation_administrationRoleARN :: Lens' StackSetOperation (Maybe Text)

The Amazon Resource Number (ARN) of the IAM role used to perform this stack set operation.

Use customized administrator roles to control which users or groups can manage specific stack sets within the same administrator account. For more information, see Define Permissions for Multiple Administrators in the CloudFormation User Guide.

stackSetOperation_action :: Lens' StackSetOperation (Maybe StackSetOperationAction)

The type of stack set operation: CREATE, UPDATE, or DELETE. Create and delete operations affect only the specified stack set instances that are associated with the specified stack set. Update operations affect both the stack set itself, as well as all associated stack set instances.

stackSetOperation_endTimestamp :: Lens' StackSetOperation (Maybe UTCTime)

The time at which the stack set operation ended, across all accounts and Regions specified. Note that this doesn't necessarily mean that the stack set operation was successful, or even attempted, in each account or Region.

stackSetOperation_creationTimestamp :: Lens' StackSetOperation (Maybe UTCTime)

The time at which the operation was initiated. Note that the creation times for the stack set operation might differ from the creation time of the individual stacks themselves. This is because CloudFormation needs to perform preparatory work for the operation, such as dispatching the work to the requested Regions, before actually creating the first stacks.

stackSetOperation_operationPreferences :: Lens' StackSetOperation (Maybe StackSetOperationPreferences)

The preferences for how CloudFormation performs this stack set operation.

stackSetOperation_operationId :: Lens' StackSetOperation (Maybe Text)

The unique ID of a stack set operation.

stackSetOperation_retainStacks :: Lens' StackSetOperation (Maybe Bool)

For stack set operations of action type DELETE, specifies whether to remove the stack instances from the specified stack set, but doesn't delete the stacks. You can't reassociate a retained stack, or add an existing, saved stack to a new stack set.

stackSetOperation_deploymentTargets :: Lens' StackSetOperation (Maybe DeploymentTargets)

Service-managed permissions

The Organizations accounts affected by the stack operation.

stackSetOperation_stackSetId :: Lens' StackSetOperation (Maybe Text)

The ID of the stack set.

stackSetOperation_executionRoleName :: Lens' StackSetOperation (Maybe Text)

The name of the IAM execution role used to create or update the stack set.

Use customized execution roles to control which stack resources users and groups can include in their stack sets.