Amazonka.CloudFormation.Types.StackSetOperationPreferences

Description

Documentation

data StackSetOperationPreferences

The user-specified preferences for how CloudFormation performs a stack set operation.

For more information on maximum concurrent accounts and failure tolerance, see Stack set operation options.

See: newStackSetOperationPreferences smart constructor.

Constructors

StackSetOperationPreferences'

Fields

regionOrder :: Maybe [Text]

The order of the Regions in where you want to perform the stack operation.

maxConcurrentCount :: Maybe Natural

The maximum number of accounts in which to perform this operation at one time. This is dependent on the value of FailureToleranceCount. MaxConcurrentCount is at most one more than the FailureToleranceCount.

Note that this setting lets you specify the maximum for operations. For large deployments, under certain circumstances the actual number of accounts acted upon concurrently may be lower due to service throttling.

Conditional: You must specify either MaxConcurrentCount or MaxConcurrentPercentage, but not both.

By default, 1 is specified.

maxConcurrentPercentage :: Maybe Natural

The maximum percentage of accounts in which to perform this operation at one time.

When calculating the number of accounts based on the specified percentage, CloudFormation rounds down to the next whole number. This is true except in cases where rounding down would result is zero. In this case, CloudFormation sets the number as one instead.

Note that this setting lets you specify the maximum for operations. For large deployments, under certain circumstances the actual number of accounts acted upon concurrently may be lower due to service throttling.

Conditional: You must specify either MaxConcurrentCount or MaxConcurrentPercentage, but not both.

By default, 1 is specified.

failureToleranceCount :: Maybe Natural

The number of accounts, per Region, for which this operation can fail before CloudFormation stops the operation in that Region. If the operation is stopped in a Region, CloudFormation doesn't attempt the operation in any subsequent Regions.

Conditional: You must specify either FailureToleranceCount or FailureTolerancePercentage (but not both).

By default, 0 is specified.

regionConcurrencyType :: Maybe RegionConcurrencyType

The concurrency type of deploying StackSets operations in Regions, could be in parallel or one Region at a time.

failureTolerancePercentage :: Maybe Natural

The percentage of accounts, per Region, for which this stack operation can fail before CloudFormation stops the operation in that Region. If the operation is stopped in a Region, CloudFormation doesn't attempt the operation in any subsequent Regions.

When calculating the number of accounts based on the specified percentage, CloudFormation rounds down to the next whole number.

Conditional: You must specify either FailureToleranceCount or FailureTolerancePercentage, but not both.

By default, 0 is specified.

newStackSetOperationPreferences :: StackSetOperationPreferences

Create a value of StackSetOperationPreferences with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:regionOrder:StackSetOperationPreferences', stackSetOperationPreferences_regionOrder - The order of the Regions in where you want to perform the stack operation.

$sel:maxConcurrentCount:StackSetOperationPreferences', stackSetOperationPreferences_maxConcurrentCount - The maximum number of accounts in which to perform this operation at one time. This is dependent on the value of FailureToleranceCount. MaxConcurrentCount is at most one more than the FailureToleranceCount.

Note that this setting lets you specify the maximum for operations. For large deployments, under certain circumstances the actual number of accounts acted upon concurrently may be lower due to service throttling.

Conditional: You must specify either MaxConcurrentCount or MaxConcurrentPercentage, but not both.

By default, 1 is specified.

$sel:maxConcurrentPercentage:StackSetOperationPreferences', stackSetOperationPreferences_maxConcurrentPercentage - The maximum percentage of accounts in which to perform this operation at one time.

When calculating the number of accounts based on the specified percentage, CloudFormation rounds down to the next whole number. This is true except in cases where rounding down would result is zero. In this case, CloudFormation sets the number as one instead.

Note that this setting lets you specify the maximum for operations. For large deployments, under certain circumstances the actual number of accounts acted upon concurrently may be lower due to service throttling.

Conditional: You must specify either MaxConcurrentCount or MaxConcurrentPercentage, but not both.

By default, 1 is specified.

$sel:failureToleranceCount:StackSetOperationPreferences', stackSetOperationPreferences_failureToleranceCount - The number of accounts, per Region, for which this operation can fail before CloudFormation stops the operation in that Region. If the operation is stopped in a Region, CloudFormation doesn't attempt the operation in any subsequent Regions.

Conditional: You must specify either FailureToleranceCount or FailureTolerancePercentage (but not both).

By default, 0 is specified.

$sel:regionConcurrencyType:StackSetOperationPreferences', stackSetOperationPreferences_regionConcurrencyType - The concurrency type of deploying StackSets operations in Regions, could be in parallel or one Region at a time.

$sel:failureTolerancePercentage:StackSetOperationPreferences', stackSetOperationPreferences_failureTolerancePercentage - The percentage of accounts, per Region, for which this stack operation can fail before CloudFormation stops the operation in that Region. If the operation is stopped in a Region, CloudFormation doesn't attempt the operation in any subsequent Regions.

When calculating the number of accounts based on the specified percentage, CloudFormation rounds down to the next whole number.

Conditional: You must specify either FailureToleranceCount or FailureTolerancePercentage, but not both.

By default, 0 is specified.

stackSetOperationPreferences_regionOrder :: Lens' StackSetOperationPreferences (Maybe [Text])

The order of the Regions in where you want to perform the stack operation.

stackSetOperationPreferences_maxConcurrentCount :: Lens' StackSetOperationPreferences (Maybe Natural)

The maximum number of accounts in which to perform this operation at one time. This is dependent on the value of FailureToleranceCount. MaxConcurrentCount is at most one more than the FailureToleranceCount.

Note that this setting lets you specify the maximum for operations. For large deployments, under certain circumstances the actual number of accounts acted upon concurrently may be lower due to service throttling.

Conditional: You must specify either MaxConcurrentCount or MaxConcurrentPercentage, but not both.

By default, 1 is specified.

stackSetOperationPreferences_maxConcurrentPercentage :: Lens' StackSetOperationPreferences (Maybe Natural)

The maximum percentage of accounts in which to perform this operation at one time.

When calculating the number of accounts based on the specified percentage, CloudFormation rounds down to the next whole number. This is true except in cases where rounding down would result is zero. In this case, CloudFormation sets the number as one instead.

Note that this setting lets you specify the maximum for operations. For large deployments, under certain circumstances the actual number of accounts acted upon concurrently may be lower due to service throttling.

Conditional: You must specify either MaxConcurrentCount or MaxConcurrentPercentage, but not both.

By default, 1 is specified.

stackSetOperationPreferences_failureToleranceCount :: Lens' StackSetOperationPreferences (Maybe Natural)

The number of accounts, per Region, for which this operation can fail before CloudFormation stops the operation in that Region. If the operation is stopped in a Region, CloudFormation doesn't attempt the operation in any subsequent Regions.

Conditional: You must specify either FailureToleranceCount or FailureTolerancePercentage (but not both).

By default, 0 is specified.

stackSetOperationPreferences_regionConcurrencyType :: Lens' StackSetOperationPreferences (Maybe RegionConcurrencyType)

The concurrency type of deploying StackSets operations in Regions, could be in parallel or one Region at a time.

stackSetOperationPreferences_failureTolerancePercentage :: Lens' StackSetOperationPreferences (Maybe Natural)

The percentage of accounts, per Region, for which this stack operation can fail before CloudFormation stops the operation in that Region. If the operation is stopped in a Region, CloudFormation doesn't attempt the operation in any subsequent Regions.

When calculating the number of accounts based on the specified percentage, CloudFormation rounds down to the next whole number.

Conditional: You must specify either FailureToleranceCount or FailureTolerancePercentage, but not both.

By default, 0 is specified.