Amazonka.CloudFormation.Types.StackSummary

Description

Documentation

data StackSummary

The StackSummary Data Type

See: newStackSummary smart constructor.

Constructors

StackSummary'

Fields

lastUpdatedTime :: Maybe ISO8601

The time the stack was last updated. This field will only be returned if the stack has been updated at least once.

rootId :: Maybe Text

For nested stacks--stacks created as resources for another stack--the stack ID of the top-level stack to which the nested stack ultimately belongs.

For more information, see Working with Nested Stacks in the CloudFormation User Guide.

stackStatusReason :: Maybe Text

Success/Failure message associated with the stack status.

templateDescription :: Maybe Text

The template description of the template used to create the stack.

driftInformation :: Maybe StackDriftInformationSummary

Summarizes information on whether a stack's actual configuration differs, or has drifted, from it's expected configuration, as defined in the stack template and any values specified as template parameters. For more information, see Detecting Unregulated Configuration Changes to Stacks and Resources.

deletionTime :: Maybe ISO8601

The time the stack was deleted.

stackId :: Maybe Text

Unique stack identifier.

parentId :: Maybe Text

For nested stacks--stacks created as resources for another stack--the stack ID of the direct parent of this stack. For the first level of nested stacks, the root stack is also the parent stack.

For more information, see Working with Nested Stacks in the CloudFormation User Guide.

stackName :: Text

The name associated with the stack.

creationTime :: ISO8601

The time the stack was created.

stackStatus :: StackStatus

The current status of the stack.

newStackSummary

Arguments

:: Text
$sel:stackName:StackSummary'

-> UTCTime
$sel:creationTime:StackSummary'

-> StackStatus
$sel:stackStatus:StackSummary'

-> StackSummary

Create a value of StackSummary with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:lastUpdatedTime:StackSummary', stackSummary_lastUpdatedTime - The time the stack was last updated. This field will only be returned if the stack has been updated at least once.

$sel:rootId:StackSummary', stackSummary_rootId - For nested stacks--stacks created as resources for another stack--the stack ID of the top-level stack to which the nested stack ultimately belongs.

For more information, see Working with Nested Stacks in the CloudFormation User Guide.

$sel:stackStatusReason:StackSummary', stackSummary_stackStatusReason - Success/Failure message associated with the stack status.

$sel:templateDescription:StackSummary', stackSummary_templateDescription - The template description of the template used to create the stack.

$sel:driftInformation:StackSummary', stackSummary_driftInformation - Summarizes information on whether a stack's actual configuration differs, or has drifted, from it's expected configuration, as defined in the stack template and any values specified as template parameters. For more information, see Detecting Unregulated Configuration Changes to Stacks and Resources.

$sel:deletionTime:StackSummary', stackSummary_deletionTime - The time the stack was deleted.

$sel:stackId:StackSummary', stackSummary_stackId - Unique stack identifier.

$sel:parentId:StackSummary', stackSummary_parentId - For nested stacks--stacks created as resources for another stack--the stack ID of the direct parent of this stack. For the first level of nested stacks, the root stack is also the parent stack.

For more information, see Working with Nested Stacks in the CloudFormation User Guide.

$sel:stackName:StackSummary', stackSummary_stackName - The name associated with the stack.

$sel:creationTime:StackSummary', stackSummary_creationTime - The time the stack was created.

$sel:stackStatus:StackSummary', stackSummary_stackStatus - The current status of the stack.

stackSummary_lastUpdatedTime :: Lens' StackSummary (Maybe UTCTime)

The time the stack was last updated. This field will only be returned if the stack has been updated at least once.

stackSummary_rootId :: Lens' StackSummary (Maybe Text)

For nested stacks--stacks created as resources for another stack--the stack ID of the top-level stack to which the nested stack ultimately belongs.

For more information, see Working with Nested Stacks in the CloudFormation User Guide.

stackSummary_stackStatusReason :: Lens' StackSummary (Maybe Text)

Success/Failure message associated with the stack status.

stackSummary_templateDescription :: Lens' StackSummary (Maybe Text)

The template description of the template used to create the stack.

stackSummary_driftInformation :: Lens' StackSummary (Maybe StackDriftInformationSummary)

Summarizes information on whether a stack's actual configuration differs, or has drifted, from it's expected configuration, as defined in the stack template and any values specified as template parameters. For more information, see Detecting Unregulated Configuration Changes to Stacks and Resources.

stackSummary_deletionTime :: Lens' StackSummary (Maybe UTCTime)

The time the stack was deleted.

stackSummary_stackId :: Lens' StackSummary (Maybe Text)

Unique stack identifier.

stackSummary_parentId :: Lens' StackSummary (Maybe Text)

For nested stacks--stacks created as resources for another stack--the stack ID of the direct parent of this stack. For the first level of nested stacks, the root stack is also the parent stack.

For more information, see Working with Nested Stacks in the CloudFormation User Guide.

stackSummary_stackName :: Lens' StackSummary Text

The name associated with the stack.

stackSummary_creationTime :: Lens' StackSummary UTCTime

The time the stack was created.

stackSummary_stackStatus :: Lens' StackSummary StackStatus

The current status of the stack.