Amazonka.CloudFormation.Types.TypeSummary

Description

Documentation

data TypeSummary

Contains summary information about the specified CloudFormation extension.

See: newTypeSummary smart constructor.

Constructors

TypeSummary'

Fields

lastUpdated :: Maybe ISO8601

When the specified extension version was registered. This applies only to:

Private extensions you have registered in your account. For more information, see RegisterType.

Public extensions you have activated in your account with auto-update specified. For more information, see ActivateType.

For all other extension types, CloudFormation returns null.

typeName :: Maybe Text

The name of the extension.

If you specified a TypeNameAlias when you activate this extension in your account and region, CloudFormation considers that alias as the type name.

publicVersionNumber :: Maybe Text

For public extensions that have been activated for this account and region, the version of the public extension to be used for CloudFormation operations in this account and region.

How you specified AutoUpdate when enabling the extension affects whether CloudFormation automatically updates the extention in this account and region when a new version is released. For more information, see Setting CloudFormation to automatically use new versions of extensions in the CloudFormation User Guide.

publisherIdentity :: Maybe IdentityProvider

The service used to verify the publisher identity.

For more information, see Registering your account to publish CloudFormation extensions in the CFN-CLI User Guide for Extension Development.

originalTypeName :: Maybe Text

For public extensions that have been activated for this account and region, the type name of the public extension.

If you specified a TypeNameAlias when enabling the extension in this account and region, CloudFormation treats that alias as the extension's type name within the account and region, not the type name of the public extension. For more information, see Specifying aliases to refer to extensions in the CloudFormation User Guide.

defaultVersionId :: Maybe Text

The ID of the default version of the extension. The default version is used when the extension version is not specified.

This applies only to private extensions you have registered in your account. For public extensions, both those provided by Amazon and published by third parties, CloudFormation returns null. For more information, see RegisterType.

To set the default version of an extension, use SetTypeDefaultVersion .

isActivated :: Maybe Bool

Whether or not the extension is activated for this account and region.

This applies only to third-party public extensions. Extensions published by Amazon are activated by default.

typeArn :: Maybe Text

The Amazon Resource Name (ARN) of the extension.

type' :: Maybe RegistryType

The kind of extension.

latestPublicVersion :: Maybe Text

For public extensions that have been activated for this account and region, the latest version of the public extension that is available. For any extensions other than activated third-arty extensions, CloudFormation returns null.

How you specified AutoUpdate when enabling the extension affects whether CloudFormation automatically updates the extention in this account and region when a new version is released. For more information, see Setting CloudFormation to automatically use new versions of extensions in the CloudFormation User Guide.

publisherId :: Maybe Text

The ID of the extension publisher, if the extension is published by a third party. Extensions published by Amazon do not return a publisher ID.

description :: Maybe Text

The description of the extension.

publisherName :: Maybe Text

The publisher name, as defined in the public profile for that publisher in the service used to verify the publisher identity.

newTypeSummary :: TypeSummary

Create a value of TypeSummary with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:lastUpdated:TypeSummary', typeSummary_lastUpdated - When the specified extension version was registered. This applies only to:

Private extensions you have registered in your account. For more information, see RegisterType.

Public extensions you have activated in your account with auto-update specified. For more information, see ActivateType.

For all other extension types, CloudFormation returns null.

$sel:typeName:TypeSummary', typeSummary_typeName - The name of the extension.

If you specified a TypeNameAlias when you activate this extension in your account and region, CloudFormation considers that alias as the type name.

$sel:publicVersionNumber:TypeSummary', typeSummary_publicVersionNumber - For public extensions that have been activated for this account and region, the version of the public extension to be used for CloudFormation operations in this account and region.

How you specified AutoUpdate when enabling the extension affects whether CloudFormation automatically updates the extention in this account and region when a new version is released. For more information, see Setting CloudFormation to automatically use new versions of extensions in the CloudFormation User Guide.

$sel:publisherIdentity:TypeSummary', typeSummary_publisherIdentity - The service used to verify the publisher identity.

For more information, see Registering your account to publish CloudFormation extensions in the CFN-CLI User Guide for Extension Development.

$sel:originalTypeName:TypeSummary', typeSummary_originalTypeName - For public extensions that have been activated for this account and region, the type name of the public extension.

If you specified a TypeNameAlias when enabling the extension in this account and region, CloudFormation treats that alias as the extension's type name within the account and region, not the type name of the public extension. For more information, see Specifying aliases to refer to extensions in the CloudFormation User Guide.

$sel:defaultVersionId:TypeSummary', typeSummary_defaultVersionId - The ID of the default version of the extension. The default version is used when the extension version is not specified.

This applies only to private extensions you have registered in your account. For public extensions, both those provided by Amazon and published by third parties, CloudFormation returns null. For more information, see RegisterType.

To set the default version of an extension, use SetTypeDefaultVersion .

$sel:isActivated:TypeSummary', typeSummary_isActivated - Whether or not the extension is activated for this account and region.

This applies only to third-party public extensions. Extensions published by Amazon are activated by default.

$sel:typeArn:TypeSummary', typeSummary_typeArn - The Amazon Resource Name (ARN) of the extension.

$sel:type':TypeSummary', typeSummary_type - The kind of extension.

$sel:latestPublicVersion:TypeSummary', typeSummary_latestPublicVersion - For public extensions that have been activated for this account and region, the latest version of the public extension that is available. For any extensions other than activated third-arty extensions, CloudFormation returns null.

How you specified AutoUpdate when enabling the extension affects whether CloudFormation automatically updates the extention in this account and region when a new version is released. For more information, see Setting CloudFormation to automatically use new versions of extensions in the CloudFormation User Guide.

$sel:publisherId:TypeSummary', typeSummary_publisherId - The ID of the extension publisher, if the extension is published by a third party. Extensions published by Amazon do not return a publisher ID.

$sel:description:TypeSummary', typeSummary_description - The description of the extension.

$sel:publisherName:TypeSummary', typeSummary_publisherName - The publisher name, as defined in the public profile for that publisher in the service used to verify the publisher identity.

typeSummary_lastUpdated :: Lens' TypeSummary (Maybe UTCTime)

When the specified extension version was registered. This applies only to:

Private extensions you have registered in your account. For more information, see RegisterType.

Public extensions you have activated in your account with auto-update specified. For more information, see ActivateType.

For all other extension types, CloudFormation returns null.

typeSummary_typeName :: Lens' TypeSummary (Maybe Text)

The name of the extension.

If you specified a TypeNameAlias when you activate this extension in your account and region, CloudFormation considers that alias as the type name.

typeSummary_publicVersionNumber :: Lens' TypeSummary (Maybe Text)

For public extensions that have been activated for this account and region, the version of the public extension to be used for CloudFormation operations in this account and region.

How you specified AutoUpdate when enabling the extension affects whether CloudFormation automatically updates the extention in this account and region when a new version is released. For more information, see Setting CloudFormation to automatically use new versions of extensions in the CloudFormation User Guide.

typeSummary_publisherIdentity :: Lens' TypeSummary (Maybe IdentityProvider)

The service used to verify the publisher identity.

For more information, see Registering your account to publish CloudFormation extensions in the CFN-CLI User Guide for Extension Development.

typeSummary_originalTypeName :: Lens' TypeSummary (Maybe Text)

For public extensions that have been activated for this account and region, the type name of the public extension.

If you specified a TypeNameAlias when enabling the extension in this account and region, CloudFormation treats that alias as the extension's type name within the account and region, not the type name of the public extension. For more information, see Specifying aliases to refer to extensions in the CloudFormation User Guide.

typeSummary_defaultVersionId :: Lens' TypeSummary (Maybe Text)

The ID of the default version of the extension. The default version is used when the extension version is not specified.

This applies only to private extensions you have registered in your account. For public extensions, both those provided by Amazon and published by third parties, CloudFormation returns null. For more information, see RegisterType.

To set the default version of an extension, use SetTypeDefaultVersion .

typeSummary_isActivated :: Lens' TypeSummary (Maybe Bool)

Whether or not the extension is activated for this account and region.

This applies only to third-party public extensions. Extensions published by Amazon are activated by default.

typeSummary_typeArn :: Lens' TypeSummary (Maybe Text)

The Amazon Resource Name (ARN) of the extension.

typeSummary_type :: Lens' TypeSummary (Maybe RegistryType)

The kind of extension.

typeSummary_latestPublicVersion :: Lens' TypeSummary (Maybe Text)

For public extensions that have been activated for this account and region, the latest version of the public extension that is available. For any extensions other than activated third-arty extensions, CloudFormation returns null.

How you specified AutoUpdate when enabling the extension affects whether CloudFormation automatically updates the extention in this account and region when a new version is released. For more information, see Setting CloudFormation to automatically use new versions of extensions in the CloudFormation User Guide.

typeSummary_publisherId :: Lens' TypeSummary (Maybe Text)

The ID of the extension publisher, if the extension is published by a third party. Extensions published by Amazon do not return a publisher ID.

typeSummary_description :: Lens' TypeSummary (Maybe Text)

The description of the extension.

typeSummary_publisherName :: Lens' TypeSummary (Maybe Text)

The publisher name, as defined in the public profile for that publisher in the service used to verify the publisher identity.