Amazonka.CloudFormation.Types.TypeVersionSummary

Description

Documentation

data TypeVersionSummary

Contains summary information about a specific version of a CloudFormation extension.

See: newTypeVersionSummary smart constructor.

Constructors

TypeVersionSummary'

Fields

versionId :: Maybe Text

The ID of a specific version of the extension. The version ID is the value at the end of the Amazon Resource Name (ARN) assigned to the extension version when it is registered.

typeName :: Maybe Text

The name of the extension.

arn :: Maybe Text

The Amazon Resource Name (ARN) of the extension version.

publicVersionNumber :: Maybe Text

For public extensions that have been activated for this account and region, the version of the public extension to be used for CloudFormation operations in this account and region. For any extensions other than activated third-arty extensions, CloudFormation returns null.

How you specified AutoUpdate when enabling the extension affects whether CloudFormation automatically updates the extention in this account and region when a new version is released. For more information, see Setting CloudFormation to automatically use new versions of extensions in the CloudFormation User Guide.

timeCreated :: Maybe ISO8601

When the version was registered.

type' :: Maybe RegistryType

The kind of extension.

isDefaultVersion :: Maybe Bool

Whether the specified extension version is set as the default version.

This applies only to private extensions you have registered in your account, and extensions published by Amazon. For public third-party extensions, whether or not they are activated in your account, CloudFormation returns null.

description :: Maybe Text

The description of the extension version.

newTypeVersionSummary :: TypeVersionSummary

Create a value of TypeVersionSummary with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:versionId:TypeVersionSummary', typeVersionSummary_versionId - The ID of a specific version of the extension. The version ID is the value at the end of the Amazon Resource Name (ARN) assigned to the extension version when it is registered.

$sel:typeName:TypeVersionSummary', typeVersionSummary_typeName - The name of the extension.

$sel:arn:TypeVersionSummary', typeVersionSummary_arn - The Amazon Resource Name (ARN) of the extension version.

$sel:publicVersionNumber:TypeVersionSummary', typeVersionSummary_publicVersionNumber - For public extensions that have been activated for this account and region, the version of the public extension to be used for CloudFormation operations in this account and region. For any extensions other than activated third-arty extensions, CloudFormation returns null.

How you specified AutoUpdate when enabling the extension affects whether CloudFormation automatically updates the extention in this account and region when a new version is released. For more information, see Setting CloudFormation to automatically use new versions of extensions in the CloudFormation User Guide.

$sel:timeCreated:TypeVersionSummary', typeVersionSummary_timeCreated - When the version was registered.

$sel:type':TypeVersionSummary', typeVersionSummary_type - The kind of extension.

$sel:isDefaultVersion:TypeVersionSummary', typeVersionSummary_isDefaultVersion - Whether the specified extension version is set as the default version.

This applies only to private extensions you have registered in your account, and extensions published by Amazon. For public third-party extensions, whether or not they are activated in your account, CloudFormation returns null.

$sel:description:TypeVersionSummary', typeVersionSummary_description - The description of the extension version.

typeVersionSummary_versionId :: Lens' TypeVersionSummary (Maybe Text)

The ID of a specific version of the extension. The version ID is the value at the end of the Amazon Resource Name (ARN) assigned to the extension version when it is registered.

typeVersionSummary_typeName :: Lens' TypeVersionSummary (Maybe Text)

The name of the extension.

typeVersionSummary_arn :: Lens' TypeVersionSummary (Maybe Text)

The Amazon Resource Name (ARN) of the extension version.

typeVersionSummary_publicVersionNumber :: Lens' TypeVersionSummary (Maybe Text)

For public extensions that have been activated for this account and region, the version of the public extension to be used for CloudFormation operations in this account and region. For any extensions other than activated third-arty extensions, CloudFormation returns null.

How you specified AutoUpdate when enabling the extension affects whether CloudFormation automatically updates the extention in this account and region when a new version is released. For more information, see Setting CloudFormation to automatically use new versions of extensions in the CloudFormation User Guide.

typeVersionSummary_timeCreated :: Lens' TypeVersionSummary (Maybe UTCTime)

When the version was registered.

typeVersionSummary_type :: Lens' TypeVersionSummary (Maybe RegistryType)

The kind of extension.

typeVersionSummary_isDefaultVersion :: Lens' TypeVersionSummary (Maybe Bool)

Whether the specified extension version is set as the default version.

This applies only to private extensions you have registered in your account, and extensions published by Amazon. For public third-party extensions, whether or not they are activated in your account, CloudFormation returns null.

typeVersionSummary_description :: Lens' TypeVersionSummary (Maybe Text)

The description of the extension version.