Amazonka.CloudFront.CreateKeyGroup

Description

Creates a key group that you can use with CloudFront signed URLs and signed cookies.

To create a key group, you must specify at least one public key for the key group. After you create a key group, you can reference it from one or more cache behaviors. When you reference a key group in a cache behavior, CloudFront requires signed URLs or signed cookies for all requests that match the cache behavior. The URLs or cookies must be signed with a private key whose corresponding public key is in the key group. The signed URL or cookie contains information about which public key CloudFront should use to verify the signature. For more information, see Serving private content in the Amazon CloudFront Developer Guide.

Creating a Request

data CreateKeyGroup

See: newCreateKeyGroup smart constructor.

Constructors

CreateKeyGroup'

Fields

keyGroupConfig :: KeyGroupConfig

A key group configuration.

newCreateKeyGroup

Arguments

:: KeyGroupConfig
$sel:keyGroupConfig:CreateKeyGroup'

-> CreateKeyGroup

Create a value of CreateKeyGroup with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:keyGroupConfig:CreateKeyGroup', createKeyGroup_keyGroupConfig - A key group configuration.

Request Lenses

createKeyGroup_keyGroupConfig :: Lens' CreateKeyGroup KeyGroupConfig

A key group configuration.

Destructuring the Response

data CreateKeyGroupResponse

See: newCreateKeyGroupResponse smart constructor.

Constructors

CreateKeyGroupResponse'

Fields

eTag :: Maybe Text

The identifier for this version of the key group.

location :: Maybe Text

The URL of the key group.

keyGroup :: Maybe KeyGroup

The key group that was just created.

httpStatus :: Int

The response's http status code.

newCreateKeyGroupResponse

Arguments

:: Int
$sel:httpStatus:CreateKeyGroupResponse'

-> CreateKeyGroupResponse

Create a value of CreateKeyGroupResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:eTag:CreateKeyGroupResponse', createKeyGroupResponse_eTag - The identifier for this version of the key group.

$sel:location:CreateKeyGroupResponse', createKeyGroupResponse_location - The URL of the key group.

$sel:keyGroup:CreateKeyGroupResponse', createKeyGroupResponse_keyGroup - The key group that was just created.

$sel:httpStatus:CreateKeyGroupResponse', createKeyGroupResponse_httpStatus - The response's http status code.

Response Lenses

createKeyGroupResponse_eTag :: Lens' CreateKeyGroupResponse (Maybe Text)

The identifier for this version of the key group.

createKeyGroupResponse_location :: Lens' CreateKeyGroupResponse (Maybe Text)

The URL of the key group.

createKeyGroupResponse_keyGroup :: Lens' CreateKeyGroupResponse (Maybe KeyGroup)

The key group that was just created.

createKeyGroupResponse_httpStatus :: Lens' CreateKeyGroupResponse Int

The response's http status code.