Amazonka.CloudFront.DeletePublicKey

Description

Remove a public key you previously added to CloudFront.

Creating a Request

data DeletePublicKey

See: newDeletePublicKey smart constructor.

Constructors

DeletePublicKey'

Fields

ifMatch :: Maybe Text

The value of the ETag header that you received when retrieving the public key identity to delete. For example: E2QWRUHAPOMQZL.

id :: Text

The ID of the public key you want to remove from CloudFront.

newDeletePublicKey

Arguments

:: Text
$sel:id:DeletePublicKey'

-> DeletePublicKey

Create a value of DeletePublicKey with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:ifMatch:DeletePublicKey', deletePublicKey_ifMatch - The value of the ETag header that you received when retrieving the public key identity to delete. For example: E2QWRUHAPOMQZL.

$sel:id:DeletePublicKey', deletePublicKey_id - The ID of the public key you want to remove from CloudFront.

Request Lenses

deletePublicKey_ifMatch :: Lens' DeletePublicKey (Maybe Text)

The value of the ETag header that you received when retrieving the public key identity to delete. For example: E2QWRUHAPOMQZL.

deletePublicKey_id :: Lens' DeletePublicKey Text

The ID of the public key you want to remove from CloudFront.

Destructuring the Response

data DeletePublicKeyResponse

See: newDeletePublicKeyResponse smart constructor.

Constructors

DeletePublicKeyResponse'

newDeletePublicKeyResponse :: DeletePublicKeyResponse

Create a value of DeletePublicKeyResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.