Amazonka.CloudFront.DescribeFunction

Description

Gets configuration information and metadata about a CloudFront function, but not the function’s code. To get a function’s code, use GetFunction.

To get configuration information and metadata about a function, you must provide the function’s name and stage. To get these values, you can use ListFunctions.

Creating a Request

data DescribeFunction

See: newDescribeFunction smart constructor.

Constructors

DescribeFunction'

Fields

stage :: Maybe FunctionStage

The function’s stage, either DEVELOPMENT or LIVE.

name :: Text

The name of the function that you are getting information about.

newDescribeFunction

Arguments

:: Text
$sel:name:DescribeFunction'

-> DescribeFunction

Create a value of DescribeFunction with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:stage:DescribeFunction', describeFunction_stage - The function’s stage, either DEVELOPMENT or LIVE.

$sel:name:DescribeFunction', describeFunction_name - The name of the function that you are getting information about.

Request Lenses

describeFunction_stage :: Lens' DescribeFunction (Maybe FunctionStage)

The function’s stage, either DEVELOPMENT or LIVE.

describeFunction_name :: Lens' DescribeFunction Text

The name of the function that you are getting information about.

Destructuring the Response

data DescribeFunctionResponse

See: newDescribeFunctionResponse smart constructor.

Constructors

DescribeFunctionResponse'

Fields

eTag :: Maybe Text

The version identifier for the current version of the CloudFront function.

functionSummary :: Maybe FunctionSummary

Contains configuration information and metadata about a CloudFront function.

httpStatus :: Int

The response's http status code.

newDescribeFunctionResponse

Arguments

:: Int
$sel:httpStatus:DescribeFunctionResponse'

-> DescribeFunctionResponse

Create a value of DescribeFunctionResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:eTag:DescribeFunctionResponse', describeFunctionResponse_eTag - The version identifier for the current version of the CloudFront function.

$sel:functionSummary:DescribeFunctionResponse', describeFunctionResponse_functionSummary - Contains configuration information and metadata about a CloudFront function.

$sel:httpStatus:DescribeFunctionResponse', describeFunctionResponse_httpStatus - The response's http status code.

Response Lenses

describeFunctionResponse_eTag :: Lens' DescribeFunctionResponse (Maybe Text)

The version identifier for the current version of the CloudFront function.

describeFunctionResponse_functionSummary :: Lens' DescribeFunctionResponse (Maybe FunctionSummary)

Contains configuration information and metadata about a CloudFront function.

describeFunctionResponse_httpStatus :: Lens' DescribeFunctionResponse Int

The response's http status code.