Amazonka.CloudFront.GetCloudFrontOriginAccessIdentityConfig

Description

Get the configuration information about an origin access identity.

Creating a Request

data GetCloudFrontOriginAccessIdentityConfig

The origin access identity's configuration information. For more information, see CloudFrontOriginAccessIdentityConfig.

See: newGetCloudFrontOriginAccessIdentityConfig smart constructor.

Constructors

GetCloudFrontOriginAccessIdentityConfig'

Fields

id :: Text

The identity's ID.

newGetCloudFrontOriginAccessIdentityConfig

Arguments

:: Text
$sel:id:GetCloudFrontOriginAccessIdentityConfig'

-> GetCloudFrontOriginAccessIdentityConfig

Create a value of GetCloudFrontOriginAccessIdentityConfig with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:id:GetCloudFrontOriginAccessIdentityConfig', getCloudFrontOriginAccessIdentityConfig_id - The identity's ID.

Request Lenses

getCloudFrontOriginAccessIdentityConfig_id :: Lens' GetCloudFrontOriginAccessIdentityConfig Text

The identity's ID.

Destructuring the Response

data GetCloudFrontOriginAccessIdentityConfigResponse

The returned result of the corresponding request.

See: newGetCloudFrontOriginAccessIdentityConfigResponse smart constructor.

Constructors

GetCloudFrontOriginAccessIdentityConfigResponse'

Fields

cloudFrontOriginAccessIdentityConfig :: Maybe CloudFrontOriginAccessIdentityConfig

The origin access identity's configuration information.

eTag :: Maybe Text

The current version of the configuration. For example: E2QWRUHAPOMQZL.

httpStatus :: Int

The response's http status code.

newGetCloudFrontOriginAccessIdentityConfigResponse

Arguments

:: Int
$sel:httpStatus:GetCloudFrontOriginAccessIdentityConfigResponse'

-> GetCloudFrontOriginAccessIdentityConfigResponse

Create a value of GetCloudFrontOriginAccessIdentityConfigResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:cloudFrontOriginAccessIdentityConfig:GetCloudFrontOriginAccessIdentityConfigResponse', getCloudFrontOriginAccessIdentityConfigResponse_cloudFrontOriginAccessIdentityConfig - The origin access identity's configuration information.

$sel:eTag:GetCloudFrontOriginAccessIdentityConfigResponse', getCloudFrontOriginAccessIdentityConfigResponse_eTag - The current version of the configuration. For example: E2QWRUHAPOMQZL.

$sel:httpStatus:GetCloudFrontOriginAccessIdentityConfigResponse', getCloudFrontOriginAccessIdentityConfigResponse_httpStatus - The response's http status code.

Response Lenses

getCloudFrontOriginAccessIdentityConfigResponse_cloudFrontOriginAccessIdentityConfig :: Lens' GetCloudFrontOriginAccessIdentityConfigResponse (Maybe CloudFrontOriginAccessIdentityConfig)

The origin access identity's configuration information.

getCloudFrontOriginAccessIdentityConfigResponse_eTag :: Lens' GetCloudFrontOriginAccessIdentityConfigResponse (Maybe Text)

The current version of the configuration. For example: E2QWRUHAPOMQZL.

getCloudFrontOriginAccessIdentityConfigResponse_httpStatus :: Lens' GetCloudFrontOriginAccessIdentityConfigResponse Int

The response's http status code.