Amazonka.CloudFront.GetDistributionConfig

Description

Get the configuration information about a distribution.

Creating a Request

data GetDistributionConfig

The request to get a distribution configuration.

See: newGetDistributionConfig smart constructor.

Constructors

GetDistributionConfig'

Fields

id :: Text

The distribution's ID. If the ID is empty, an empty distribution configuration is returned.

newGetDistributionConfig

Arguments

:: Text
$sel:id:GetDistributionConfig'

-> GetDistributionConfig

Create a value of GetDistributionConfig with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:id:GetDistributionConfig', getDistributionConfig_id - The distribution's ID. If the ID is empty, an empty distribution configuration is returned.

Request Lenses

getDistributionConfig_id :: Lens' GetDistributionConfig Text

The distribution's ID. If the ID is empty, an empty distribution configuration is returned.

Destructuring the Response

data GetDistributionConfigResponse

The returned result of the corresponding request.

See: newGetDistributionConfigResponse smart constructor.

Constructors

GetDistributionConfigResponse'

Fields

eTag :: Maybe Text

The current version of the configuration. For example: E2QWRUHAPOMQZL.

distributionConfig :: Maybe DistributionConfig

The distribution's configuration information.

httpStatus :: Int

The response's http status code.

newGetDistributionConfigResponse

Arguments

:: Int
$sel:httpStatus:GetDistributionConfigResponse'

-> GetDistributionConfigResponse

Create a value of GetDistributionConfigResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:eTag:GetDistributionConfigResponse', getDistributionConfigResponse_eTag - The current version of the configuration. For example: E2QWRUHAPOMQZL.

$sel:distributionConfig:GetDistributionConfigResponse', getDistributionConfigResponse_distributionConfig - The distribution's configuration information.

$sel:httpStatus:GetDistributionConfigResponse', getDistributionConfigResponse_httpStatus - The response's http status code.

Response Lenses

getDistributionConfigResponse_eTag :: Lens' GetDistributionConfigResponse (Maybe Text)

The current version of the configuration. For example: E2QWRUHAPOMQZL.

getDistributionConfigResponse_distributionConfig :: Lens' GetDistributionConfigResponse (Maybe DistributionConfig)

The distribution's configuration information.

getDistributionConfigResponse_httpStatus :: Lens' GetDistributionConfigResponse Int

The response's http status code.