Amazonka.CloudFront.GetInvalidation

Description

Get the information about an invalidation.

Creating a Request

data GetInvalidation

The request to get an invalidation's information.

See: newGetInvalidation smart constructor.

Constructors

GetInvalidation'

Fields

distributionId :: Text

The distribution's ID.

id :: Text

The identifier for the invalidation request, for example, IDFDVBD632BHDS5.

newGetInvalidation

Arguments

:: Text
$sel:distributionId:GetInvalidation'

-> Text
$sel:id:GetInvalidation'

-> GetInvalidation

Create a value of GetInvalidation with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:distributionId:GetInvalidation', getInvalidation_distributionId - The distribution's ID.

$sel:id:GetInvalidation', getInvalidation_id - The identifier for the invalidation request, for example, IDFDVBD632BHDS5.

Request Lenses

getInvalidation_distributionId :: Lens' GetInvalidation Text

The distribution's ID.

getInvalidation_id :: Lens' GetInvalidation Text

The identifier for the invalidation request, for example, IDFDVBD632BHDS5.

Destructuring the Response

data GetInvalidationResponse

The returned result of the corresponding request.

See: newGetInvalidationResponse smart constructor.

Constructors

GetInvalidationResponse'

Fields

invalidation :: Maybe Invalidation

The invalidation's information. For more information, see Invalidation Complex Type.

httpStatus :: Int

The response's http status code.

newGetInvalidationResponse

Arguments

:: Int
$sel:httpStatus:GetInvalidationResponse'

-> GetInvalidationResponse

Create a value of GetInvalidationResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:invalidation:GetInvalidationResponse', getInvalidationResponse_invalidation - The invalidation's information. For more information, see Invalidation Complex Type.

$sel:httpStatus:GetInvalidationResponse', getInvalidationResponse_httpStatus - The response's http status code.

Response Lenses

getInvalidationResponse_invalidation :: Lens' GetInvalidationResponse (Maybe Invalidation)

The invalidation's information. For more information, see Invalidation Complex Type.

getInvalidationResponse_httpStatus :: Lens' GetInvalidationResponse Int

The response's http status code.