Amazonka.CloudFront.GetRealtimeLogConfig

Description

Gets a real-time log configuration.

To get a real-time log configuration, you can provide the configuration’s name or its Amazon Resource Name (ARN). You must provide at least one. If you provide both, CloudFront uses the name to identify the real-time log configuration to get.

Creating a Request

data GetRealtimeLogConfig

See: newGetRealtimeLogConfig smart constructor.

Constructors

GetRealtimeLogConfig'

Fields

arn :: Maybe Text

The Amazon Resource Name (ARN) of the real-time log configuration to get.

name :: Maybe Text

The name of the real-time log configuration to get.

newGetRealtimeLogConfig :: GetRealtimeLogConfig

Create a value of GetRealtimeLogConfig with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:arn:GetRealtimeLogConfig', getRealtimeLogConfig_arn - The Amazon Resource Name (ARN) of the real-time log configuration to get.

$sel:name:GetRealtimeLogConfig', getRealtimeLogConfig_name - The name of the real-time log configuration to get.

Request Lenses

getRealtimeLogConfig_arn :: Lens' GetRealtimeLogConfig (Maybe Text)

The Amazon Resource Name (ARN) of the real-time log configuration to get.

getRealtimeLogConfig_name :: Lens' GetRealtimeLogConfig (Maybe Text)

The name of the real-time log configuration to get.

Destructuring the Response

data GetRealtimeLogConfigResponse

See: newGetRealtimeLogConfigResponse smart constructor.

Constructors

GetRealtimeLogConfigResponse'

Fields

realtimeLogConfig :: Maybe RealtimeLogConfig

A real-time log configuration.

httpStatus :: Int

The response's http status code.

newGetRealtimeLogConfigResponse

Arguments

:: Int
$sel:httpStatus:GetRealtimeLogConfigResponse'

-> GetRealtimeLogConfigResponse

Create a value of GetRealtimeLogConfigResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:realtimeLogConfig:GetRealtimeLogConfigResponse', getRealtimeLogConfigResponse_realtimeLogConfig - A real-time log configuration.

$sel:httpStatus:GetRealtimeLogConfigResponse', getRealtimeLogConfigResponse_httpStatus - The response's http status code.

Response Lenses

getRealtimeLogConfigResponse_realtimeLogConfig :: Lens' GetRealtimeLogConfigResponse (Maybe RealtimeLogConfig)

A real-time log configuration.

getRealtimeLogConfigResponse_httpStatus :: Lens' GetRealtimeLogConfigResponse Int

The response's http status code.