Amazonka.CloudFront.ListDistributionsByKeyGroup

Description

Gets a list of distribution IDs for distributions that have a cache behavior that references the specified key group.

You can optionally specify the maximum number of items to receive in the response. If the total number of items in the list exceeds the maximum that you specify, or the default maximum, the response is paginated. To get the next page of items, send a subsequent request that specifies the NextMarker value from the current response as the Marker value in the subsequent request.

Creating a Request

data ListDistributionsByKeyGroup

See: newListDistributionsByKeyGroup smart constructor.

Constructors

ListDistributionsByKeyGroup'

Fields

marker :: Maybe Text

Use this field when paginating results to indicate where to begin in your list of distribution IDs. The response includes distribution IDs in the list that occur after the marker. To get the next page of the list, set this field’s value to the value of NextMarker from the current page’s response.

maxItems :: Maybe Text

The maximum number of distribution IDs that you want in the response.

keyGroupId :: Text

The ID of the key group whose associated distribution IDs you are listing.

newListDistributionsByKeyGroup

Arguments

:: Text
$sel:keyGroupId:ListDistributionsByKeyGroup'

-> ListDistributionsByKeyGroup

Create a value of ListDistributionsByKeyGroup with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:marker:ListDistributionsByKeyGroup', listDistributionsByKeyGroup_marker - Use this field when paginating results to indicate where to begin in your list of distribution IDs. The response includes distribution IDs in the list that occur after the marker. To get the next page of the list, set this field’s value to the value of NextMarker from the current page’s response.

$sel:maxItems:ListDistributionsByKeyGroup', listDistributionsByKeyGroup_maxItems - The maximum number of distribution IDs that you want in the response.

$sel:keyGroupId:ListDistributionsByKeyGroup', listDistributionsByKeyGroup_keyGroupId - The ID of the key group whose associated distribution IDs you are listing.

Request Lenses

listDistributionsByKeyGroup_marker :: Lens' ListDistributionsByKeyGroup (Maybe Text)

Use this field when paginating results to indicate where to begin in your list of distribution IDs. The response includes distribution IDs in the list that occur after the marker. To get the next page of the list, set this field’s value to the value of NextMarker from the current page’s response.

listDistributionsByKeyGroup_maxItems :: Lens' ListDistributionsByKeyGroup (Maybe Text)

The maximum number of distribution IDs that you want in the response.

listDistributionsByKeyGroup_keyGroupId :: Lens' ListDistributionsByKeyGroup Text

The ID of the key group whose associated distribution IDs you are listing.

Destructuring the Response

data ListDistributionsByKeyGroupResponse

See: newListDistributionsByKeyGroupResponse smart constructor.

Constructors

ListDistributionsByKeyGroupResponse'

Fields

distributionIdList :: Maybe DistributionIdList

httpStatus :: Int

The response's http status code.

newListDistributionsByKeyGroupResponse

Arguments

:: Int
$sel:httpStatus:ListDistributionsByKeyGroupResponse'

-> ListDistributionsByKeyGroupResponse

Create a value of ListDistributionsByKeyGroupResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:distributionIdList:ListDistributionsByKeyGroupResponse', listDistributionsByKeyGroupResponse_distributionIdList - Undocumented member.

$sel:httpStatus:ListDistributionsByKeyGroupResponse', listDistributionsByKeyGroupResponse_httpStatus - The response's http status code.

Response Lenses

listDistributionsByKeyGroupResponse_distributionIdList :: Lens' ListDistributionsByKeyGroupResponse (Maybe DistributionIdList)

Undocumented member.

listDistributionsByKeyGroupResponse_httpStatus :: Lens' ListDistributionsByKeyGroupResponse Int

The response's http status code.