Amazonka.CloudFront.Types.CachePolicyCookiesConfig

Description

Documentation

data CachePolicyCookiesConfig

An object that determines whether any cookies in viewer requests (and if so, which cookies) are included in the cache key and automatically included in requests that CloudFront sends to the origin.

See: newCachePolicyCookiesConfig smart constructor.

Constructors

CachePolicyCookiesConfig'

Fields

cookies :: Maybe CookieNames

cookieBehavior :: CachePolicyCookieBehavior

Determines whether any cookies in viewer requests are included in the cache key and automatically included in requests that CloudFront sends to the origin. Valid values are:

none – Cookies in viewer requests are not included in the cache key and are not automatically included in requests that CloudFront sends to the origin. Even when this field is set to none, any cookies that are listed in an OriginRequestPolicy are included in origin requests.

whitelist – The cookies in viewer requests that are listed in the CookieNames type are included in the cache key and automatically included in requests that CloudFront sends to the origin.

allExcept – All cookies in viewer requests that are not listed in the CookieNames type are included in the cache key and automatically included in requests that CloudFront sends to the origin.

all – All cookies in viewer requests are included in the cache key and are automatically included in requests that CloudFront sends to the origin.

newCachePolicyCookiesConfig

Arguments

:: CachePolicyCookieBehavior
$sel:cookieBehavior:CachePolicyCookiesConfig'

-> CachePolicyCookiesConfig

Create a value of CachePolicyCookiesConfig with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:cookies:CachePolicyCookiesConfig', cachePolicyCookiesConfig_cookies - Undocumented member.

$sel:cookieBehavior:CachePolicyCookiesConfig', cachePolicyCookiesConfig_cookieBehavior - Determines whether any cookies in viewer requests are included in the cache key and automatically included in requests that CloudFront sends to the origin. Valid values are:

none – Cookies in viewer requests are not included in the cache key and are not automatically included in requests that CloudFront sends to the origin. Even when this field is set to none, any cookies that are listed in an OriginRequestPolicy are included in origin requests.

whitelist – The cookies in viewer requests that are listed in the CookieNames type are included in the cache key and automatically included in requests that CloudFront sends to the origin.

allExcept – All cookies in viewer requests that are not listed in the CookieNames type are included in the cache key and automatically included in requests that CloudFront sends to the origin.

all – All cookies in viewer requests are included in the cache key and are automatically included in requests that CloudFront sends to the origin.

cachePolicyCookiesConfig_cookies :: Lens' CachePolicyCookiesConfig (Maybe CookieNames)

Undocumented member.

cachePolicyCookiesConfig_cookieBehavior :: Lens' CachePolicyCookiesConfig CachePolicyCookieBehavior

Determines whether any cookies in viewer requests are included in the cache key and automatically included in requests that CloudFront sends to the origin. Valid values are:

none – Cookies in viewer requests are not included in the cache key and are not automatically included in requests that CloudFront sends to the origin. Even when this field is set to none, any cookies that are listed in an OriginRequestPolicy are included in origin requests.

whitelist – The cookies in viewer requests that are listed in the CookieNames type are included in the cache key and automatically included in requests that CloudFront sends to the origin.

allExcept – All cookies in viewer requests that are not listed in the CookieNames type are included in the cache key and automatically included in requests that CloudFront sends to the origin.

all – All cookies in viewer requests are included in the cache key and are automatically included in requests that CloudFront sends to the origin.