Amazonka.CloudFront.Types.CachePolicyList

Description

Documentation

data CachePolicyList

A list of cache policies.

See: newCachePolicyList smart constructor.

Constructors

CachePolicyList'

Fields

items :: Maybe [CachePolicySummary]

Contains the cache policies in the list.

nextMarker :: Maybe Text

If there are more items in the list than are in this response, this element is present. It contains the value that you should use in the Marker field of a subsequent request to continue listing cache policies where you left off.

maxItems :: Int

The maximum number of cache policies requested.

quantity :: Int

The total number of cache policies returned in the response.

newCachePolicyList

Arguments

:: Int
$sel:maxItems:CachePolicyList'

-> Int
$sel:quantity:CachePolicyList'

-> CachePolicyList

Create a value of CachePolicyList with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:items:CachePolicyList', cachePolicyList_items - Contains the cache policies in the list.

$sel:nextMarker:CachePolicyList', cachePolicyList_nextMarker - If there are more items in the list than are in this response, this element is present. It contains the value that you should use in the Marker field of a subsequent request to continue listing cache policies where you left off.

$sel:maxItems:CachePolicyList', cachePolicyList_maxItems - The maximum number of cache policies requested.

$sel:quantity:CachePolicyList', cachePolicyList_quantity - The total number of cache policies returned in the response.

cachePolicyList_items :: Lens' CachePolicyList (Maybe [CachePolicySummary])

Contains the cache policies in the list.

cachePolicyList_nextMarker :: Lens' CachePolicyList (Maybe Text)

If there are more items in the list than are in this response, this element is present. It contains the value that you should use in the Marker field of a subsequent request to continue listing cache policies where you left off.

cachePolicyList_maxItems :: Lens' CachePolicyList Int

The maximum number of cache policies requested.

cachePolicyList_quantity :: Lens' CachePolicyList Int

The total number of cache policies returned in the response.