Amazonka.CloudFront.Types.DistributionConfigWithTags

Description

Documentation

data DistributionConfigWithTags

A distribution Configuration and a list of tags to be associated with the distribution.

See: newDistributionConfigWithTags smart constructor.

Constructors

DistributionConfigWithTags'

Fields

distributionConfig :: DistributionConfig

A distribution configuration.

tags :: Tags

A complex type that contains zero or more Tag elements.

newDistributionConfigWithTags

Arguments

:: DistributionConfig
$sel:distributionConfig:DistributionConfigWithTags'

-> Tags
$sel:tags:DistributionConfigWithTags'

-> DistributionConfigWithTags

Create a value of DistributionConfigWithTags with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:distributionConfig:DistributionConfigWithTags', distributionConfigWithTags_distributionConfig - A distribution configuration.

$sel:tags:DistributionConfigWithTags', distributionConfigWithTags_tags - A complex type that contains zero or more Tag elements.

distributionConfigWithTags_distributionConfig :: Lens' DistributionConfigWithTags DistributionConfig

A distribution configuration.

distributionConfigWithTags_tags :: Lens' DistributionConfigWithTags Tags

A complex type that contains zero or more Tag elements.