Amazonka.CloudFront.Types.EncryptionEntity

Description

Documentation

data EncryptionEntity

Complex data type for field-level encryption profiles that includes the encryption key and field pattern specifications.

See: newEncryptionEntity smart constructor.

Constructors

EncryptionEntity'

Fields

publicKeyId :: Text

The public key associated with a set of field-level encryption patterns, to be used when encrypting the fields that match the patterns.

providerId :: Text

The provider associated with the public key being used for encryption. This value must also be provided with the private key for applications to be able to decrypt data.

fieldPatterns :: FieldPatterns

Field patterns in a field-level encryption content type profile specify the fields that you want to be encrypted. You can provide the full field name, or any beginning characters followed by a wildcard (*). You can't overlap field patterns. For example, you can't have both ABC* and AB*. Note that field patterns are case-sensitive.

newEncryptionEntity

Arguments

:: Text
$sel:publicKeyId:EncryptionEntity'

-> Text
$sel:providerId:EncryptionEntity'

-> FieldPatterns
$sel:fieldPatterns:EncryptionEntity'

-> EncryptionEntity

Create a value of EncryptionEntity with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:publicKeyId:EncryptionEntity', encryptionEntity_publicKeyId - The public key associated with a set of field-level encryption patterns, to be used when encrypting the fields that match the patterns.

$sel:providerId:EncryptionEntity', encryptionEntity_providerId - The provider associated with the public key being used for encryption. This value must also be provided with the private key for applications to be able to decrypt data.

$sel:fieldPatterns:EncryptionEntity', encryptionEntity_fieldPatterns - Field patterns in a field-level encryption content type profile specify the fields that you want to be encrypted. You can provide the full field name, or any beginning characters followed by a wildcard (*). You can't overlap field patterns. For example, you can't have both ABC* and AB*. Note that field patterns are case-sensitive.

encryptionEntity_publicKeyId :: Lens' EncryptionEntity Text

The public key associated with a set of field-level encryption patterns, to be used when encrypting the fields that match the patterns.

encryptionEntity_providerId :: Lens' EncryptionEntity Text

The provider associated with the public key being used for encryption. This value must also be provided with the private key for applications to be able to decrypt data.

encryptionEntity_fieldPatterns :: Lens' EncryptionEntity FieldPatterns

Field patterns in a field-level encryption content type profile specify the fields that you want to be encrypted. You can provide the full field name, or any beginning characters followed by a wildcard (*). You can't overlap field patterns. For example, you can't have both ABC* and AB*. Note that field patterns are case-sensitive.