Amazonka.CloudFront.Types.EndPoint

Description

Documentation

data EndPoint

Contains information about the Amazon Kinesis data stream where you are sending real-time log data in a real-time log configuration.

See: newEndPoint smart constructor.

Constructors

EndPoint'

Fields

kinesisStreamConfig :: Maybe KinesisStreamConfig

Contains information about the Amazon Kinesis data stream where you are sending real-time log data.

streamType :: Text

The type of data stream where you are sending real-time log data. The only valid value is Kinesis.

newEndPoint

Arguments

:: Text
$sel:streamType:EndPoint'

-> EndPoint

Create a value of EndPoint with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:kinesisStreamConfig:EndPoint', endPoint_kinesisStreamConfig - Contains information about the Amazon Kinesis data stream where you are sending real-time log data.

$sel:streamType:EndPoint', endPoint_streamType - The type of data stream where you are sending real-time log data. The only valid value is Kinesis.

endPoint_kinesisStreamConfig :: Lens' EndPoint (Maybe KinesisStreamConfig)

Contains information about the Amazon Kinesis data stream where you are sending real-time log data.

endPoint_streamType :: Lens' EndPoint Text

The type of data stream where you are sending real-time log data. The only valid value is Kinesis.