Amazonka.CloudFront.Types.FieldLevelEncryptionProfileConfig

Description

Documentation

data FieldLevelEncryptionProfileConfig

A complex data type of profiles for the field-level encryption.

See: newFieldLevelEncryptionProfileConfig smart constructor.

Constructors

FieldLevelEncryptionProfileConfig'

Fields

comment :: Maybe Text

An optional comment for the field-level encryption profile. The comment cannot be longer than 128 characters.

name :: Text

Profile name for the field-level encryption profile.

callerReference :: Text

A unique number that ensures that the request can't be replayed.

encryptionEntities :: EncryptionEntities

A complex data type of encryption entities for the field-level encryption profile that include the public key ID, provider, and field patterns for specifying which fields to encrypt with this key.

newFieldLevelEncryptionProfileConfig

Arguments

:: Text
$sel:name:FieldLevelEncryptionProfileConfig'

-> Text
$sel:callerReference:FieldLevelEncryptionProfileConfig'

-> EncryptionEntities
$sel:encryptionEntities:FieldLevelEncryptionProfileConfig'

-> FieldLevelEncryptionProfileConfig

Create a value of FieldLevelEncryptionProfileConfig with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:comment:FieldLevelEncryptionProfileConfig', fieldLevelEncryptionProfileConfig_comment - An optional comment for the field-level encryption profile. The comment cannot be longer than 128 characters.

$sel:name:FieldLevelEncryptionProfileConfig', fieldLevelEncryptionProfileConfig_name - Profile name for the field-level encryption profile.

$sel:callerReference:FieldLevelEncryptionProfileConfig', fieldLevelEncryptionProfileConfig_callerReference - A unique number that ensures that the request can't be replayed.

$sel:encryptionEntities:FieldLevelEncryptionProfileConfig', fieldLevelEncryptionProfileConfig_encryptionEntities - A complex data type of encryption entities for the field-level encryption profile that include the public key ID, provider, and field patterns for specifying which fields to encrypt with this key.

fieldLevelEncryptionProfileConfig_comment :: Lens' FieldLevelEncryptionProfileConfig (Maybe Text)

An optional comment for the field-level encryption profile. The comment cannot be longer than 128 characters.

fieldLevelEncryptionProfileConfig_name :: Lens' FieldLevelEncryptionProfileConfig Text

Profile name for the field-level encryption profile.

fieldLevelEncryptionProfileConfig_callerReference :: Lens' FieldLevelEncryptionProfileConfig Text

A unique number that ensures that the request can't be replayed.

fieldLevelEncryptionProfileConfig_encryptionEntities :: Lens' FieldLevelEncryptionProfileConfig EncryptionEntities

A complex data type of encryption entities for the field-level encryption profile that include the public key ID, provider, and field patterns for specifying which fields to encrypt with this key.