Amazonka.CloudFront.Types.FieldLevelEncryptionProfileSummary

Description

Documentation

data FieldLevelEncryptionProfileSummary

The field-level encryption profile summary.

See: newFieldLevelEncryptionProfileSummary smart constructor.

Constructors

FieldLevelEncryptionProfileSummary'

Fields

comment :: Maybe Text

An optional comment for the field-level encryption profile summary. The comment cannot be longer than 128 characters.

id :: Text

ID for the field-level encryption profile summary.

lastModifiedTime :: ISO8601

The time when the the field-level encryption profile summary was last updated.

name :: Text

Name for the field-level encryption profile summary.

encryptionEntities :: EncryptionEntities

A complex data type of encryption entities for the field-level encryption profile that include the public key ID, provider, and field patterns for specifying which fields to encrypt with this key.

newFieldLevelEncryptionProfileSummary

Arguments

:: Text
$sel:id:FieldLevelEncryptionProfileSummary'

-> UTCTime
$sel:lastModifiedTime:FieldLevelEncryptionProfileSummary'

-> Text
$sel:name:FieldLevelEncryptionProfileSummary'

-> EncryptionEntities
$sel:encryptionEntities:FieldLevelEncryptionProfileSummary'

-> FieldLevelEncryptionProfileSummary

Create a value of FieldLevelEncryptionProfileSummary with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:comment:FieldLevelEncryptionProfileSummary', fieldLevelEncryptionProfileSummary_comment - An optional comment for the field-level encryption profile summary. The comment cannot be longer than 128 characters.

$sel:id:FieldLevelEncryptionProfileSummary', fieldLevelEncryptionProfileSummary_id - ID for the field-level encryption profile summary.

$sel:lastModifiedTime:FieldLevelEncryptionProfileSummary', fieldLevelEncryptionProfileSummary_lastModifiedTime - The time when the the field-level encryption profile summary was last updated.

$sel:name:FieldLevelEncryptionProfileSummary', fieldLevelEncryptionProfileSummary_name - Name for the field-level encryption profile summary.

$sel:encryptionEntities:FieldLevelEncryptionProfileSummary', fieldLevelEncryptionProfileSummary_encryptionEntities - A complex data type of encryption entities for the field-level encryption profile that include the public key ID, provider, and field patterns for specifying which fields to encrypt with this key.

fieldLevelEncryptionProfileSummary_comment :: Lens' FieldLevelEncryptionProfileSummary (Maybe Text)

An optional comment for the field-level encryption profile summary. The comment cannot be longer than 128 characters.

fieldLevelEncryptionProfileSummary_id :: Lens' FieldLevelEncryptionProfileSummary Text

ID for the field-level encryption profile summary.

fieldLevelEncryptionProfileSummary_lastModifiedTime :: Lens' FieldLevelEncryptionProfileSummary UTCTime

The time when the the field-level encryption profile summary was last updated.

fieldLevelEncryptionProfileSummary_name :: Lens' FieldLevelEncryptionProfileSummary Text

Name for the field-level encryption profile summary.

fieldLevelEncryptionProfileSummary_encryptionEntities :: Lens' FieldLevelEncryptionProfileSummary EncryptionEntities

A complex data type of encryption entities for the field-level encryption profile that include the public key ID, provider, and field patterns for specifying which fields to encrypt with this key.