Amazonka.CloudFront.Types.FunctionList

Description

Documentation

data FunctionList

A list of CloudFront functions.

See: newFunctionList smart constructor.

Constructors

FunctionList'

Fields

items :: Maybe [FunctionSummary]

Contains the functions in the list.

nextMarker :: Maybe Text

If there are more items in the list than are in this response, this element is present. It contains the value that you should use in the Marker field of a subsequent request to continue listing functions where you left off.

maxItems :: Int

The maximum number of functions requested.

quantity :: Int

The number of functions returned in the response.

newFunctionList

Arguments

:: Int
$sel:maxItems:FunctionList'

-> Int
$sel:quantity:FunctionList'

-> FunctionList

Create a value of FunctionList with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:items:FunctionList', functionList_items - Contains the functions in the list.

$sel:nextMarker:FunctionList', functionList_nextMarker - If there are more items in the list than are in this response, this element is present. It contains the value that you should use in the Marker field of a subsequent request to continue listing functions where you left off.

$sel:maxItems:FunctionList', functionList_maxItems - The maximum number of functions requested.

$sel:quantity:FunctionList', functionList_quantity - The number of functions returned in the response.

functionList_items :: Lens' FunctionList (Maybe [FunctionSummary])

Contains the functions in the list.

functionList_nextMarker :: Lens' FunctionList (Maybe Text)

If there are more items in the list than are in this response, this element is present. It contains the value that you should use in the Marker field of a subsequent request to continue listing functions where you left off.

functionList_maxItems :: Lens' FunctionList Int

The maximum number of functions requested.

functionList_quantity :: Lens' FunctionList Int

The number of functions returned in the response.