Amazonka.CloudFront.Types.LambdaFunctionAssociations

Description

Documentation

data LambdaFunctionAssociations

A complex type that specifies a list of Lambda@Edge functions associations for a cache behavior.

If you want to invoke one or more Lambda@Edge functions triggered by requests that match the PathPattern of the cache behavior, specify the applicable values for Quantity and Items. Note that there can be up to 4 LambdaFunctionAssociation items in this list (one for each possible value of EventType) and each EventType can be associated with only one function.

If you don't want to invoke any Lambda@Edge functions for the requests that match PathPattern, specify 0 for Quantity and omit Items.

See: newLambdaFunctionAssociations smart constructor.

Constructors

LambdaFunctionAssociations'

Fields

items :: Maybe [LambdaFunctionAssociation]

Optional: A complex type that contains LambdaFunctionAssociation items for this cache behavior. If Quantity is 0, you can omit Items.

quantity :: Int

The number of Lambda@Edge function associations for this cache behavior.

newLambdaFunctionAssociations

Arguments

:: Int
$sel:quantity:LambdaFunctionAssociations'

-> LambdaFunctionAssociations

Create a value of LambdaFunctionAssociations with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:items:LambdaFunctionAssociations', lambdaFunctionAssociations_items - Optional: A complex type that contains LambdaFunctionAssociation items for this cache behavior. If Quantity is 0, you can omit Items.

$sel:quantity:LambdaFunctionAssociations', lambdaFunctionAssociations_quantity - The number of Lambda@Edge function associations for this cache behavior.

lambdaFunctionAssociations_items :: Lens' LambdaFunctionAssociations (Maybe [LambdaFunctionAssociation])

Optional: A complex type that contains LambdaFunctionAssociation items for this cache behavior. If Quantity is 0, you can omit Items.

lambdaFunctionAssociations_quantity :: Lens' LambdaFunctionAssociations Int

The number of Lambda@Edge function associations for this cache behavior.