Amazonka.CloudFront.Types.LoggingConfig

Description

Documentation

data LoggingConfig

A complex type that controls whether access logs are written for the distribution.

See: newLoggingConfig smart constructor.

Constructors

LoggingConfig'

Fields

enabled :: Bool

Specifies whether you want CloudFront to save access logs to an Amazon S3 bucket. If you don't want to enable logging when you create a distribution or if you want to disable logging for an existing distribution, specify false for Enabled, and specify empty Bucket and Prefix elements. If you specify false for Enabled but you specify values for Bucket, prefix, and IncludeCookies, the values are automatically deleted.

includeCookies :: Bool

Specifies whether you want CloudFront to include cookies in access logs, specify true for IncludeCookies. If you choose to include cookies in logs, CloudFront logs all cookies regardless of how you configure the cache behaviors for this distribution. If you don't want to include cookies when you create a distribution or if you want to disable include cookies for an existing distribution, specify false for IncludeCookies.

bucket :: Text

The Amazon S3 bucket to store the access logs in, for example, myawslogbucket.s3.amazonaws.com.

prefix :: Text

An optional string that you want CloudFront to prefix to the access log filenames for this distribution, for example, myprefix/. If you want to enable logging, but you don't want to specify a prefix, you still must include an empty Prefix element in the Logging element.

newLoggingConfig

Arguments

:: Bool
$sel:enabled:LoggingConfig'

-> Bool
$sel:includeCookies:LoggingConfig'

-> Text
$sel:bucket:LoggingConfig'

-> Text
$sel:prefix:LoggingConfig'

-> LoggingConfig

Create a value of LoggingConfig with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:enabled:LoggingConfig', loggingConfig_enabled - Specifies whether you want CloudFront to save access logs to an Amazon S3 bucket. If you don't want to enable logging when you create a distribution or if you want to disable logging for an existing distribution, specify false for Enabled, and specify empty Bucket and Prefix elements. If you specify false for Enabled but you specify values for Bucket, prefix, and IncludeCookies, the values are automatically deleted.

$sel:includeCookies:LoggingConfig', loggingConfig_includeCookies - Specifies whether you want CloudFront to include cookies in access logs, specify true for IncludeCookies. If you choose to include cookies in logs, CloudFront logs all cookies regardless of how you configure the cache behaviors for this distribution. If you don't want to include cookies when you create a distribution or if you want to disable include cookies for an existing distribution, specify false for IncludeCookies.

$sel:bucket:LoggingConfig', loggingConfig_bucket - The Amazon S3 bucket to store the access logs in, for example, myawslogbucket.s3.amazonaws.com.

$sel:prefix:LoggingConfig', loggingConfig_prefix - An optional string that you want CloudFront to prefix to the access log filenames for this distribution, for example, myprefix/. If you want to enable logging, but you don't want to specify a prefix, you still must include an empty Prefix element in the Logging element.

loggingConfig_enabled :: Lens' LoggingConfig Bool

Specifies whether you want CloudFront to save access logs to an Amazon S3 bucket. If you don't want to enable logging when you create a distribution or if you want to disable logging for an existing distribution, specify false for Enabled, and specify empty Bucket and Prefix elements. If you specify false for Enabled but you specify values for Bucket, prefix, and IncludeCookies, the values are automatically deleted.

loggingConfig_includeCookies :: Lens' LoggingConfig Bool

Specifies whether you want CloudFront to include cookies in access logs, specify true for IncludeCookies. If you choose to include cookies in logs, CloudFront logs all cookies regardless of how you configure the cache behaviors for this distribution. If you don't want to include cookies when you create a distribution or if you want to disable include cookies for an existing distribution, specify false for IncludeCookies.

loggingConfig_bucket :: Lens' LoggingConfig Text

The Amazon S3 bucket to store the access logs in, for example, myawslogbucket.s3.amazonaws.com.

loggingConfig_prefix :: Lens' LoggingConfig Text

An optional string that you want CloudFront to prefix to the access log filenames for this distribution, for example, myprefix/. If you want to enable logging, but you don't want to specify a prefix, you still must include an empty Prefix element in the Logging element.