Amazonka.CloudFront.Types.OriginRequestPolicyConfig

Description

Documentation

data OriginRequestPolicyConfig

An origin request policy configuration.

This configuration determines the values that CloudFront includes in requests that it sends to the origin. Each request that CloudFront sends to the origin includes the following:

The request body and the URL path (without the domain name) from the viewer request.

The headers that CloudFront automatically includes in every origin request, including Host, User-Agent, and X-Amz-Cf-Id.

All HTTP headers, cookies, and URL query strings that are specified in the cache policy or the origin request policy. These can include items from the viewer request and, in the case of headers, additional ones that are added by CloudFront.

CloudFront sends a request when it can’t find an object in its cache that matches the request. If you want to send values to the origin and also include them in the cache key, use CachePolicy.

See: newOriginRequestPolicyConfig smart constructor.

Constructors

OriginRequestPolicyConfig'

Fields

comment :: Maybe Text

A comment to describe the origin request policy. The comment cannot be longer than 128 characters.

name :: Text

A unique name to identify the origin request policy.

headersConfig :: OriginRequestPolicyHeadersConfig

The HTTP headers to include in origin requests. These can include headers from viewer requests and additional headers added by CloudFront.

cookiesConfig :: OriginRequestPolicyCookiesConfig

The cookies from viewer requests to include in origin requests.

queryStringsConfig :: OriginRequestPolicyQueryStringsConfig

The URL query strings from viewer requests to include in origin requests.

newOriginRequestPolicyConfig

Arguments

:: Text
$sel:name:OriginRequestPolicyConfig'

-> OriginRequestPolicyHeadersConfig
$sel:headersConfig:OriginRequestPolicyConfig'

-> OriginRequestPolicyCookiesConfig
$sel:cookiesConfig:OriginRequestPolicyConfig'

-> OriginRequestPolicyQueryStringsConfig
$sel:queryStringsConfig:OriginRequestPolicyConfig'

-> OriginRequestPolicyConfig

Create a value of OriginRequestPolicyConfig with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:comment:OriginRequestPolicyConfig', originRequestPolicyConfig_comment - A comment to describe the origin request policy. The comment cannot be longer than 128 characters.

$sel:name:OriginRequestPolicyConfig', originRequestPolicyConfig_name - A unique name to identify the origin request policy.

$sel:headersConfig:OriginRequestPolicyConfig', originRequestPolicyConfig_headersConfig - The HTTP headers to include in origin requests. These can include headers from viewer requests and additional headers added by CloudFront.

$sel:cookiesConfig:OriginRequestPolicyConfig', originRequestPolicyConfig_cookiesConfig - The cookies from viewer requests to include in origin requests.

$sel:queryStringsConfig:OriginRequestPolicyConfig', originRequestPolicyConfig_queryStringsConfig - The URL query strings from viewer requests to include in origin requests.

originRequestPolicyConfig_comment :: Lens' OriginRequestPolicyConfig (Maybe Text)

A comment to describe the origin request policy. The comment cannot be longer than 128 characters.

originRequestPolicyConfig_name :: Lens' OriginRequestPolicyConfig Text

A unique name to identify the origin request policy.

originRequestPolicyConfig_headersConfig :: Lens' OriginRequestPolicyConfig OriginRequestPolicyHeadersConfig

The HTTP headers to include in origin requests. These can include headers from viewer requests and additional headers added by CloudFront.

originRequestPolicyConfig_cookiesConfig :: Lens' OriginRequestPolicyConfig OriginRequestPolicyCookiesConfig

The cookies from viewer requests to include in origin requests.

originRequestPolicyConfig_queryStringsConfig :: Lens' OriginRequestPolicyConfig OriginRequestPolicyQueryStringsConfig

The URL query strings from viewer requests to include in origin requests.