Amazonka.CloudFront.Types.OriginRequestPolicyHeadersConfig

Description

Documentation

data OriginRequestPolicyHeadersConfig

An object that determines whether any HTTP headers (and if so, which headers) are included in requests that CloudFront sends to the origin.

See: newOriginRequestPolicyHeadersConfig smart constructor.

Constructors

OriginRequestPolicyHeadersConfig'

Fields

headers :: Maybe Headers

headerBehavior :: OriginRequestPolicyHeaderBehavior

Determines whether any HTTP headers are included in requests that CloudFront sends to the origin. Valid values are:

none – HTTP headers are not included in requests that CloudFront sends to the origin. Even when this field is set to none, any headers that are listed in a CachePolicy are included in origin requests.

whitelist – The HTTP headers that are listed in the Headers type are included in requests that CloudFront sends to the origin.

allViewer – All HTTP headers in viewer requests are included in requests that CloudFront sends to the origin.

allViewerAndWhitelistCloudFront – All HTTP headers in viewer requests and the additional CloudFront headers that are listed in the Headers type are included in requests that CloudFront sends to the origin. The additional headers are added by CloudFront.

newOriginRequestPolicyHeadersConfig

Arguments

:: OriginRequestPolicyHeaderBehavior
$sel:headerBehavior:OriginRequestPolicyHeadersConfig'

-> OriginRequestPolicyHeadersConfig

Create a value of OriginRequestPolicyHeadersConfig with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:headers:OriginRequestPolicyHeadersConfig', originRequestPolicyHeadersConfig_headers - Undocumented member.

$sel:headerBehavior:OriginRequestPolicyHeadersConfig', originRequestPolicyHeadersConfig_headerBehavior - Determines whether any HTTP headers are included in requests that CloudFront sends to the origin. Valid values are:

none – HTTP headers are not included in requests that CloudFront sends to the origin. Even when this field is set to none, any headers that are listed in a CachePolicy are included in origin requests.

whitelist – The HTTP headers that are listed in the Headers type are included in requests that CloudFront sends to the origin.

allViewer – All HTTP headers in viewer requests are included in requests that CloudFront sends to the origin.

allViewerAndWhitelistCloudFront – All HTTP headers in viewer requests and the additional CloudFront headers that are listed in the Headers type are included in requests that CloudFront sends to the origin. The additional headers are added by CloudFront.

originRequestPolicyHeadersConfig_headers :: Lens' OriginRequestPolicyHeadersConfig (Maybe Headers)

Undocumented member.

originRequestPolicyHeadersConfig_headerBehavior :: Lens' OriginRequestPolicyHeadersConfig OriginRequestPolicyHeaderBehavior

Determines whether any HTTP headers are included in requests that CloudFront sends to the origin. Valid values are:

none – HTTP headers are not included in requests that CloudFront sends to the origin. Even when this field is set to none, any headers that are listed in a CachePolicy are included in origin requests.

whitelist – The HTTP headers that are listed in the Headers type are included in requests that CloudFront sends to the origin.

allViewer – All HTTP headers in viewer requests are included in requests that CloudFront sends to the origin.

allViewerAndWhitelistCloudFront – All HTTP headers in viewer requests and the additional CloudFront headers that are listed in the Headers type are included in requests that CloudFront sends to the origin. The additional headers are added by CloudFront.