Amazonka.CloudFront.Types.ParametersInCacheKeyAndForwardedToOrigin

Description

Documentation

data ParametersInCacheKeyAndForwardedToOrigin

This object determines the values that CloudFront includes in the cache key. These values can include HTTP headers, cookies, and URL query strings. CloudFront uses the cache key to find an object in its cache that it can return to the viewer.

The headers, cookies, and query strings that are included in the cache key are automatically included in requests that CloudFront sends to the origin. CloudFront sends a request when it can’t find an object in its cache that matches the request’s cache key. If you want to send values to the origin but not include them in the cache key, use OriginRequestPolicy.

See: newParametersInCacheKeyAndForwardedToOrigin smart constructor.

Constructors

ParametersInCacheKeyAndForwardedToOrigin'

Fields

enableAcceptEncodingBrotli :: Maybe Bool

A flag that can affect whether the Accept-Encoding HTTP header is included in the cache key and included in requests that CloudFront sends to the origin.

This field is related to the EnableAcceptEncodingGzip field. If one or both of these fields is true and the viewer request includes the Accept-Encoding header, then CloudFront does the following:

Normalizes the value of the viewer’s Accept-Encoding header

Includes the normalized header in the cache key

Includes the normalized header in the request to the origin, if a request is necessary

For more information, see Compression support in the Amazon CloudFront Developer Guide.

If you set this value to true, and this cache behavior also has an origin request policy attached, do not include the Accept-Encoding header in the origin request policy. CloudFront always includes the Accept-Encoding header in origin requests when the value of this field is true, so including this header in an origin request policy has no effect.

If both of these fields are false, then CloudFront treats the Accept-Encoding header the same as any other HTTP header in the viewer request. By default, it’s not included in the cache key and it’s not included in origin requests. In this case, you can manually add Accept-Encoding to the headers whitelist like any other HTTP header.

enableAcceptEncodingGzip :: Bool

A flag that can affect whether the Accept-Encoding HTTP header is included in the cache key and included in requests that CloudFront sends to the origin.

This field is related to the EnableAcceptEncodingBrotli field. If one or both of these fields is true and the viewer request includes the Accept-Encoding header, then CloudFront does the following:

Normalizes the value of the viewer’s Accept-Encoding header

Includes the normalized header in the cache key

Includes the normalized header in the request to the origin, if a request is necessary

For more information, see Compression support in the Amazon CloudFront Developer Guide.

If you set this value to true, and this cache behavior also has an origin request policy attached, do not include the Accept-Encoding header in the origin request policy. CloudFront always includes the Accept-Encoding header in origin requests when the value of this field is true, so including this header in an origin request policy has no effect.

If both of these fields are false, then CloudFront treats the Accept-Encoding header the same as any other HTTP header in the viewer request. By default, it’s not included in the cache key and it’s not included in origin requests. In this case, you can manually add Accept-Encoding to the headers whitelist like any other HTTP header.

headersConfig :: CachePolicyHeadersConfig

An object that determines whether any HTTP headers (and if so, which headers) are included in the cache key and automatically included in requests that CloudFront sends to the origin.

cookiesConfig :: CachePolicyCookiesConfig

An object that determines whether any cookies in viewer requests (and if so, which cookies) are included in the cache key and automatically included in requests that CloudFront sends to the origin.

queryStringsConfig :: CachePolicyQueryStringsConfig

An object that determines whether any URL query strings in viewer requests (and if so, which query strings) are included in the cache key and automatically included in requests that CloudFront sends to the origin.

newParametersInCacheKeyAndForwardedToOrigin

Arguments

:: Bool
$sel:enableAcceptEncodingGzip:ParametersInCacheKeyAndForwardedToOrigin'

-> CachePolicyHeadersConfig
$sel:headersConfig:ParametersInCacheKeyAndForwardedToOrigin'

-> CachePolicyCookiesConfig
$sel:cookiesConfig:ParametersInCacheKeyAndForwardedToOrigin'

-> CachePolicyQueryStringsConfig
$sel:queryStringsConfig:ParametersInCacheKeyAndForwardedToOrigin'

-> ParametersInCacheKeyAndForwardedToOrigin

Create a value of ParametersInCacheKeyAndForwardedToOrigin with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:enableAcceptEncodingBrotli:ParametersInCacheKeyAndForwardedToOrigin', parametersInCacheKeyAndForwardedToOrigin_enableAcceptEncodingBrotli - A flag that can affect whether the Accept-Encoding HTTP header is included in the cache key and included in requests that CloudFront sends to the origin.

This field is related to the EnableAcceptEncodingGzip field. If one or both of these fields is true and the viewer request includes the Accept-Encoding header, then CloudFront does the following:

Normalizes the value of the viewer’s Accept-Encoding header

Includes the normalized header in the cache key

Includes the normalized header in the request to the origin, if a request is necessary

For more information, see Compression support in the Amazon CloudFront Developer Guide.

If you set this value to true, and this cache behavior also has an origin request policy attached, do not include the Accept-Encoding header in the origin request policy. CloudFront always includes the Accept-Encoding header in origin requests when the value of this field is true, so including this header in an origin request policy has no effect.

If both of these fields are false, then CloudFront treats the Accept-Encoding header the same as any other HTTP header in the viewer request. By default, it’s not included in the cache key and it’s not included in origin requests. In this case, you can manually add Accept-Encoding to the headers whitelist like any other HTTP header.

$sel:enableAcceptEncodingGzip:ParametersInCacheKeyAndForwardedToOrigin', parametersInCacheKeyAndForwardedToOrigin_enableAcceptEncodingGzip - A flag that can affect whether the Accept-Encoding HTTP header is included in the cache key and included in requests that CloudFront sends to the origin.

This field is related to the EnableAcceptEncodingBrotli field. If one or both of these fields is true and the viewer request includes the Accept-Encoding header, then CloudFront does the following:

Normalizes the value of the viewer’s Accept-Encoding header

Includes the normalized header in the cache key

Includes the normalized header in the request to the origin, if a request is necessary

For more information, see Compression support in the Amazon CloudFront Developer Guide.

If you set this value to true, and this cache behavior also has an origin request policy attached, do not include the Accept-Encoding header in the origin request policy. CloudFront always includes the Accept-Encoding header in origin requests when the value of this field is true, so including this header in an origin request policy has no effect.

If both of these fields are false, then CloudFront treats the Accept-Encoding header the same as any other HTTP header in the viewer request. By default, it’s not included in the cache key and it’s not included in origin requests. In this case, you can manually add Accept-Encoding to the headers whitelist like any other HTTP header.

$sel:headersConfig:ParametersInCacheKeyAndForwardedToOrigin', parametersInCacheKeyAndForwardedToOrigin_headersConfig - An object that determines whether any HTTP headers (and if so, which headers) are included in the cache key and automatically included in requests that CloudFront sends to the origin.

$sel:cookiesConfig:ParametersInCacheKeyAndForwardedToOrigin', parametersInCacheKeyAndForwardedToOrigin_cookiesConfig - An object that determines whether any cookies in viewer requests (and if so, which cookies) are included in the cache key and automatically included in requests that CloudFront sends to the origin.

$sel:queryStringsConfig:ParametersInCacheKeyAndForwardedToOrigin', parametersInCacheKeyAndForwardedToOrigin_queryStringsConfig - An object that determines whether any URL query strings in viewer requests (and if so, which query strings) are included in the cache key and automatically included in requests that CloudFront sends to the origin.

parametersInCacheKeyAndForwardedToOrigin_enableAcceptEncodingBrotli :: Lens' ParametersInCacheKeyAndForwardedToOrigin (Maybe Bool)

A flag that can affect whether the Accept-Encoding HTTP header is included in the cache key and included in requests that CloudFront sends to the origin.

This field is related to the EnableAcceptEncodingGzip field. If one or both of these fields is true and the viewer request includes the Accept-Encoding header, then CloudFront does the following:

Normalizes the value of the viewer’s Accept-Encoding header

Includes the normalized header in the cache key

Includes the normalized header in the request to the origin, if a request is necessary

For more information, see Compression support in the Amazon CloudFront Developer Guide.

If you set this value to true, and this cache behavior also has an origin request policy attached, do not include the Accept-Encoding header in the origin request policy. CloudFront always includes the Accept-Encoding header in origin requests when the value of this field is true, so including this header in an origin request policy has no effect.

If both of these fields are false, then CloudFront treats the Accept-Encoding header the same as any other HTTP header in the viewer request. By default, it’s not included in the cache key and it’s not included in origin requests. In this case, you can manually add Accept-Encoding to the headers whitelist like any other HTTP header.

parametersInCacheKeyAndForwardedToOrigin_enableAcceptEncodingGzip :: Lens' ParametersInCacheKeyAndForwardedToOrigin Bool

A flag that can affect whether the Accept-Encoding HTTP header is included in the cache key and included in requests that CloudFront sends to the origin.

This field is related to the EnableAcceptEncodingBrotli field. If one or both of these fields is true and the viewer request includes the Accept-Encoding header, then CloudFront does the following:

Normalizes the value of the viewer’s Accept-Encoding header

Includes the normalized header in the cache key

Includes the normalized header in the request to the origin, if a request is necessary

For more information, see Compression support in the Amazon CloudFront Developer Guide.

If you set this value to true, and this cache behavior also has an origin request policy attached, do not include the Accept-Encoding header in the origin request policy. CloudFront always includes the Accept-Encoding header in origin requests when the value of this field is true, so including this header in an origin request policy has no effect.

If both of these fields are false, then CloudFront treats the Accept-Encoding header the same as any other HTTP header in the viewer request. By default, it’s not included in the cache key and it’s not included in origin requests. In this case, you can manually add Accept-Encoding to the headers whitelist like any other HTTP header.

parametersInCacheKeyAndForwardedToOrigin_headersConfig :: Lens' ParametersInCacheKeyAndForwardedToOrigin CachePolicyHeadersConfig

An object that determines whether any HTTP headers (and if so, which headers) are included in the cache key and automatically included in requests that CloudFront sends to the origin.

parametersInCacheKeyAndForwardedToOrigin_cookiesConfig :: Lens' ParametersInCacheKeyAndForwardedToOrigin CachePolicyCookiesConfig

An object that determines whether any cookies in viewer requests (and if so, which cookies) are included in the cache key and automatically included in requests that CloudFront sends to the origin.

parametersInCacheKeyAndForwardedToOrigin_queryStringsConfig :: Lens' ParametersInCacheKeyAndForwardedToOrigin CachePolicyQueryStringsConfig

An object that determines whether any URL query strings in viewer requests (and if so, which query strings) are included in the cache key and automatically included in requests that CloudFront sends to the origin.