Amazonka.CloudFront.Types.PublicKey

Description

Documentation

data PublicKey

A public key that you can use with signed URLs and signed cookies, or with field-level encryption.

See: newPublicKey smart constructor.

Constructors

PublicKey'

Fields

id :: Text

The identifier of the public key.

createdTime :: ISO8601

The date and time when the public key was uploaded.

publicKeyConfig :: PublicKeyConfig

Configuration information about a public key that you can use with signed URLs and signed cookies, or with field-level encryption.

newPublicKey

Arguments

:: Text
$sel:id:PublicKey'

-> UTCTime
$sel:createdTime:PublicKey'

-> PublicKeyConfig
$sel:publicKeyConfig:PublicKey'

-> PublicKey

Create a value of PublicKey with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:id:PublicKey', publicKey_id - The identifier of the public key.

$sel:createdTime:PublicKey', publicKey_createdTime - The date and time when the public key was uploaded.

$sel:publicKeyConfig:PublicKey', publicKey_publicKeyConfig - Configuration information about a public key that you can use with signed URLs and signed cookies, or with field-level encryption.

publicKey_id :: Lens' PublicKey Text

The identifier of the public key.

publicKey_createdTime :: Lens' PublicKey UTCTime

The date and time when the public key was uploaded.

publicKey_publicKeyConfig :: Lens' PublicKey PublicKeyConfig

Configuration information about a public key that you can use with signed URLs and signed cookies, or with field-level encryption.