Amazonka.CloudFront.Types.RealtimeLogConfig

Description

Documentation

data RealtimeLogConfig

A real-time log configuration.

See: newRealtimeLogConfig smart constructor.

Constructors

RealtimeLogConfig'

Fields

arn :: Text

The Amazon Resource Name (ARN) of this real-time log configuration.

name :: Text

The unique name of this real-time log configuration.

samplingRate :: Integer

The sampling rate for this real-time log configuration. The sampling rate determines the percentage of viewer requests that are represented in the real-time log data. The sampling rate is an integer between 1 and 100, inclusive.

endPoints :: [EndPoint]

Contains information about the Amazon Kinesis data stream where you are sending real-time log data for this real-time log configuration.

fields :: [Text]

A list of fields that are included in each real-time log record. In an API response, the fields are provided in the same order in which they are sent to the Amazon Kinesis data stream.

For more information about fields, see Real-time log configuration fields in the Amazon CloudFront Developer Guide.

newRealtimeLogConfig

Arguments

:: Text
$sel:arn:RealtimeLogConfig'

-> Text
$sel:name:RealtimeLogConfig'

-> Integer
$sel:samplingRate:RealtimeLogConfig'

-> RealtimeLogConfig

Create a value of RealtimeLogConfig with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:arn:RealtimeLogConfig', realtimeLogConfig_arn - The Amazon Resource Name (ARN) of this real-time log configuration.

$sel:name:RealtimeLogConfig', realtimeLogConfig_name - The unique name of this real-time log configuration.

$sel:samplingRate:RealtimeLogConfig', realtimeLogConfig_samplingRate - The sampling rate for this real-time log configuration. The sampling rate determines the percentage of viewer requests that are represented in the real-time log data. The sampling rate is an integer between 1 and 100, inclusive.

$sel:endPoints:RealtimeLogConfig', realtimeLogConfig_endPoints - Contains information about the Amazon Kinesis data stream where you are sending real-time log data for this real-time log configuration.

$sel:fields:RealtimeLogConfig', realtimeLogConfig_fields - A list of fields that are included in each real-time log record. In an API response, the fields are provided in the same order in which they are sent to the Amazon Kinesis data stream.

For more information about fields, see Real-time log configuration fields in the Amazon CloudFront Developer Guide.

realtimeLogConfig_arn :: Lens' RealtimeLogConfig Text

The Amazon Resource Name (ARN) of this real-time log configuration.

realtimeLogConfig_name :: Lens' RealtimeLogConfig Text

The unique name of this real-time log configuration.

realtimeLogConfig_samplingRate :: Lens' RealtimeLogConfig Integer

The sampling rate for this real-time log configuration. The sampling rate determines the percentage of viewer requests that are represented in the real-time log data. The sampling rate is an integer between 1 and 100, inclusive.

realtimeLogConfig_endPoints :: Lens' RealtimeLogConfig [EndPoint]

Contains information about the Amazon Kinesis data stream where you are sending real-time log data for this real-time log configuration.

realtimeLogConfig_fields :: Lens' RealtimeLogConfig [Text]

A list of fields that are included in each real-time log record. In an API response, the fields are provided in the same order in which they are sent to the Amazon Kinesis data stream.

For more information about fields, see Real-time log configuration fields in the Amazon CloudFront Developer Guide.