Amazonka.CloudFront.Types.StreamingDistribution

Description

Documentation

data StreamingDistribution

A streaming distribution tells CloudFront where you want RTMP content to be delivered from, and the details about how to track and manage content delivery.

See: newStreamingDistribution smart constructor.

Constructors

StreamingDistribution'

Fields

lastModifiedTime :: Maybe ISO8601

The date and time that the distribution was last modified.

id :: Text

The identifier for the RTMP distribution. For example: EGTXBD79EXAMPLE.

arn :: Text

The ARN (Amazon Resource Name) for the distribution. For example: arn:aws:cloudfront::123456789012:distribution/EDFDVBD632BHDS5, where 123456789012 is your account ID.

status :: Text

The current status of the RTMP distribution. When the status is Deployed, the distribution's information is propagated to all CloudFront edge locations.

domainName :: Text

The domain name that corresponds to the streaming distribution, for example, s5c39gqb8ow64r.cloudfront.net.

activeTrustedSigners :: ActiveTrustedSigners

A complex type that lists the accounts, if any, that you included in the TrustedSigners complex type for this distribution. These are the accounts that you want to allow to create signed URLs for private content.

The Signer complex type lists the account number of the trusted signer or self if the signer is the account that created the distribution. The Signer element also includes the IDs of any active CloudFront key pairs that are associated with the trusted signer's account. If no KeyPairId element appears for a Signer, that signer can't create signed URLs.

For more information, see Serving Private Content through CloudFront in the Amazon CloudFront Developer Guide.

streamingDistributionConfig :: StreamingDistributionConfig

The current configuration information for the RTMP distribution.

newStreamingDistribution

Arguments

:: Text
$sel:id:StreamingDistribution'

-> Text
$sel:arn:StreamingDistribution'

-> Text
$sel:status:StreamingDistribution'

-> Text
$sel:domainName:StreamingDistribution'

-> ActiveTrustedSigners
$sel:activeTrustedSigners:StreamingDistribution'

-> StreamingDistributionConfig
$sel:streamingDistributionConfig:StreamingDistribution'

-> StreamingDistribution

Create a value of StreamingDistribution with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:lastModifiedTime:StreamingDistribution', streamingDistribution_lastModifiedTime - The date and time that the distribution was last modified.

$sel:id:StreamingDistribution', streamingDistribution_id - The identifier for the RTMP distribution. For example: EGTXBD79EXAMPLE.

$sel:arn:StreamingDistribution', streamingDistribution_arn - The ARN (Amazon Resource Name) for the distribution. For example: arn:aws:cloudfront::123456789012:distribution/EDFDVBD632BHDS5, where 123456789012 is your account ID.

$sel:status:StreamingDistribution', streamingDistribution_status - The current status of the RTMP distribution. When the status is Deployed, the distribution's information is propagated to all CloudFront edge locations.

$sel:domainName:StreamingDistribution', streamingDistribution_domainName - The domain name that corresponds to the streaming distribution, for example, s5c39gqb8ow64r.cloudfront.net.

$sel:activeTrustedSigners:StreamingDistribution', streamingDistribution_activeTrustedSigners - A complex type that lists the accounts, if any, that you included in the TrustedSigners complex type for this distribution. These are the accounts that you want to allow to create signed URLs for private content.

The Signer complex type lists the account number of the trusted signer or self if the signer is the account that created the distribution. The Signer element also includes the IDs of any active CloudFront key pairs that are associated with the trusted signer's account. If no KeyPairId element appears for a Signer, that signer can't create signed URLs.

For more information, see Serving Private Content through CloudFront in the Amazon CloudFront Developer Guide.

$sel:streamingDistributionConfig:StreamingDistribution', streamingDistribution_streamingDistributionConfig - The current configuration information for the RTMP distribution.

streamingDistribution_lastModifiedTime :: Lens' StreamingDistribution (Maybe UTCTime)

The date and time that the distribution was last modified.

streamingDistribution_id :: Lens' StreamingDistribution Text

The identifier for the RTMP distribution. For example: EGTXBD79EXAMPLE.

streamingDistribution_arn :: Lens' StreamingDistribution Text

The ARN (Amazon Resource Name) for the distribution. For example: arn:aws:cloudfront::123456789012:distribution/EDFDVBD632BHDS5, where 123456789012 is your account ID.

streamingDistribution_status :: Lens' StreamingDistribution Text

The current status of the RTMP distribution. When the status is Deployed, the distribution's information is propagated to all CloudFront edge locations.

streamingDistribution_domainName :: Lens' StreamingDistribution Text

The domain name that corresponds to the streaming distribution, for example, s5c39gqb8ow64r.cloudfront.net.

streamingDistribution_activeTrustedSigners :: Lens' StreamingDistribution ActiveTrustedSigners

A complex type that lists the accounts, if any, that you included in the TrustedSigners complex type for this distribution. These are the accounts that you want to allow to create signed URLs for private content.

The Signer complex type lists the account number of the trusted signer or self if the signer is the account that created the distribution. The Signer element also includes the IDs of any active CloudFront key pairs that are associated with the trusted signer's account. If no KeyPairId element appears for a Signer, that signer can't create signed URLs.

For more information, see Serving Private Content through CloudFront in the Amazon CloudFront Developer Guide.

streamingDistribution_streamingDistributionConfig :: Lens' StreamingDistribution StreamingDistributionConfig

The current configuration information for the RTMP distribution.