Amazonka.CloudFront.UpdateDistribution

Description

Updates the configuration for a web distribution.

When you update a distribution, there are more required fields than when you create a distribution. When you update your distribution by using this API action, follow the steps here to get the current configuration and then make your updates, to make sure that you include all of the required fields. To view a summary, see Required Fields for Create Distribution and Update Distribution in the Amazon CloudFront Developer Guide.

The update process includes getting the current distribution configuration, updating the XML document that is returned to make your changes, and then submitting an UpdateDistribution request to make the updates.

For information about updating a distribution using the CloudFront console instead, see Creating a Distribution in the Amazon CloudFront Developer Guide.

To update a web distribution using the CloudFront API

Submit a GetDistributionConfig request to get the current configuration and an Etag header for the distribution.

If you update the distribution again, you must get a new Etag header.

Update the XML document that was returned in the response to your GetDistributionConfig request to include your changes.

When you edit the XML file, be aware of the following:

You must strip out the ETag parameter that is returned.

Additional fields are required when you update a distribution. There may be fields included in the XML file for features that you haven't configured for your distribution. This is expected and required to successfully update the distribution.

You can't change the value of CallerReference. If you try to change this value, CloudFront returns an IllegalUpdate error.

The new configuration replaces the existing configuration; the values that you specify in an UpdateDistribution request are not merged into your existing configuration. When you add, delete, or replace values in an element that allows multiple values (for example, CNAME), you must specify all of the values that you want to appear in the updated distribution. In addition, you must update the corresponding Quantity element.

Submit an UpdateDistribution request to update the configuration for your distribution:

In the request body, include the XML document that you updated in Step 2. The request body must include an XML document with a DistributionConfig element.

Set the value of the HTTP If-Match header to the value of the ETag header that CloudFront returned when you submitted the GetDistributionConfig request in Step 1.

Review the response to the UpdateDistribution request to confirm that the configuration was successfully updated.

Optional: Submit a GetDistribution request to confirm that your changes have propagated. When propagation is complete, the value of Status is Deployed.

Creating a Request

data UpdateDistribution

The request to update a distribution.

See: newUpdateDistribution smart constructor.

Constructors

UpdateDistribution'

Fields

ifMatch :: Maybe Text

The value of the ETag header that you received when retrieving the distribution's configuration. For example: E2QWRUHAPOMQZL.

distributionConfig :: DistributionConfig

The distribution's configuration information.

id :: Text

The distribution's id.

newUpdateDistribution

Arguments

:: DistributionConfig
$sel:distributionConfig:UpdateDistribution'

-> Text
$sel:id:UpdateDistribution'

-> UpdateDistribution

Create a value of UpdateDistribution with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:ifMatch:UpdateDistribution', updateDistribution_ifMatch - The value of the ETag header that you received when retrieving the distribution's configuration. For example: E2QWRUHAPOMQZL.

$sel:distributionConfig:UpdateDistribution', updateDistribution_distributionConfig - The distribution's configuration information.

$sel:id:UpdateDistribution', updateDistribution_id - The distribution's id.

Request Lenses

updateDistribution_ifMatch :: Lens' UpdateDistribution (Maybe Text)

The value of the ETag header that you received when retrieving the distribution's configuration. For example: E2QWRUHAPOMQZL.

updateDistribution_distributionConfig :: Lens' UpdateDistribution DistributionConfig

The distribution's configuration information.

updateDistribution_id :: Lens' UpdateDistribution Text

The distribution's id.

Destructuring the Response

data UpdateDistributionResponse

The returned result of the corresponding request.

See: newUpdateDistributionResponse smart constructor.

Constructors

UpdateDistributionResponse'

Fields

eTag :: Maybe Text

The current version of the configuration. For example: E2QWRUHAPOMQZL.

distribution :: Maybe Distribution

The distribution's information.

httpStatus :: Int

The response's http status code.

newUpdateDistributionResponse

Arguments

:: Int
$sel:httpStatus:UpdateDistributionResponse'

-> UpdateDistributionResponse

Create a value of UpdateDistributionResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:eTag:UpdateDistributionResponse', updateDistributionResponse_eTag - The current version of the configuration. For example: E2QWRUHAPOMQZL.

$sel:distribution:UpdateDistributionResponse', updateDistributionResponse_distribution - The distribution's information.

$sel:httpStatus:UpdateDistributionResponse', updateDistributionResponse_httpStatus - The response's http status code.

Response Lenses

updateDistributionResponse_eTag :: Lens' UpdateDistributionResponse (Maybe Text)

The current version of the configuration. For example: E2QWRUHAPOMQZL.

updateDistributionResponse_distribution :: Lens' UpdateDistributionResponse (Maybe Distribution)

The distribution's information.

updateDistributionResponse_httpStatus :: Lens' UpdateDistributionResponse Int

The response's http status code.