Amazonka.CloudFront.UpdateOriginRequestPolicy

Description

Updates an origin request policy configuration.

When you update an origin request policy configuration, all the fields are updated with the values provided in the request. You cannot update some fields independent of others. To update an origin request policy configuration:

Use GetOriginRequestPolicyConfig to get the current configuration.

Locally modify the fields in the origin request policy configuration that you want to update.

Call UpdateOriginRequestPolicy by providing the entire origin request policy configuration, including the fields that you modified and those that you didn’t.

Creating a Request

data UpdateOriginRequestPolicy

See: newUpdateOriginRequestPolicy smart constructor.

Constructors

UpdateOriginRequestPolicy'

Fields

ifMatch :: Maybe Text

The version of the origin request policy that you are updating. The version is returned in the origin request policy’s ETag field in the response to GetOriginRequestPolicyConfig.

originRequestPolicyConfig :: OriginRequestPolicyConfig

An origin request policy configuration.

id :: Text

The unique identifier for the origin request policy that you are updating. The identifier is returned in a cache behavior’s OriginRequestPolicyId field in the response to GetDistributionConfig.

newUpdateOriginRequestPolicy

Arguments

:: OriginRequestPolicyConfig
$sel:originRequestPolicyConfig:UpdateOriginRequestPolicy'

-> Text
$sel:id:UpdateOriginRequestPolicy'

-> UpdateOriginRequestPolicy

Create a value of UpdateOriginRequestPolicy with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:ifMatch:UpdateOriginRequestPolicy', updateOriginRequestPolicy_ifMatch - The version of the origin request policy that you are updating. The version is returned in the origin request policy’s ETag field in the response to GetOriginRequestPolicyConfig.

$sel:originRequestPolicyConfig:UpdateOriginRequestPolicy', updateOriginRequestPolicy_originRequestPolicyConfig - An origin request policy configuration.

$sel:id:UpdateOriginRequestPolicy', updateOriginRequestPolicy_id - The unique identifier for the origin request policy that you are updating. The identifier is returned in a cache behavior’s OriginRequestPolicyId field in the response to GetDistributionConfig.

Request Lenses

updateOriginRequestPolicy_ifMatch :: Lens' UpdateOriginRequestPolicy (Maybe Text)

The version of the origin request policy that you are updating. The version is returned in the origin request policy’s ETag field in the response to GetOriginRequestPolicyConfig.

updateOriginRequestPolicy_originRequestPolicyConfig :: Lens' UpdateOriginRequestPolicy OriginRequestPolicyConfig

An origin request policy configuration.

updateOriginRequestPolicy_id :: Lens' UpdateOriginRequestPolicy Text

The unique identifier for the origin request policy that you are updating. The identifier is returned in a cache behavior’s OriginRequestPolicyId field in the response to GetDistributionConfig.

Destructuring the Response

data UpdateOriginRequestPolicyResponse

See: newUpdateOriginRequestPolicyResponse smart constructor.

Constructors

UpdateOriginRequestPolicyResponse'

Fields

eTag :: Maybe Text

The current version of the origin request policy.

originRequestPolicy :: Maybe OriginRequestPolicy

An origin request policy.

httpStatus :: Int

The response's http status code.

newUpdateOriginRequestPolicyResponse

Arguments

:: Int
$sel:httpStatus:UpdateOriginRequestPolicyResponse'

-> UpdateOriginRequestPolicyResponse

Create a value of UpdateOriginRequestPolicyResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:eTag:UpdateOriginRequestPolicyResponse', updateOriginRequestPolicyResponse_eTag - The current version of the origin request policy.

$sel:originRequestPolicy:UpdateOriginRequestPolicyResponse', updateOriginRequestPolicyResponse_originRequestPolicy - An origin request policy.

$sel:httpStatus:UpdateOriginRequestPolicyResponse', updateOriginRequestPolicyResponse_httpStatus - The response's http status code.

Response Lenses

updateOriginRequestPolicyResponse_eTag :: Lens' UpdateOriginRequestPolicyResponse (Maybe Text)

The current version of the origin request policy.

updateOriginRequestPolicyResponse_originRequestPolicy :: Lens' UpdateOriginRequestPolicyResponse (Maybe OriginRequestPolicy)

An origin request policy.

updateOriginRequestPolicyResponse_httpStatus :: Lens' UpdateOriginRequestPolicyResponse Int

The response's http status code.