Amazonka.CloudFront.UpdatePublicKey

Description

Update public key information. Note that the only value you can change is the comment.

Creating a Request

data UpdatePublicKey

See: newUpdatePublicKey smart constructor.

Constructors

UpdatePublicKey'

Fields

ifMatch :: Maybe Text

The value of the ETag header that you received when retrieving the public key to update. For example: E2QWRUHAPOMQZL.

publicKeyConfig :: PublicKeyConfig

A public key configuration.

id :: Text

The identifier of the public key that you are updating.

newUpdatePublicKey

Arguments

:: PublicKeyConfig
$sel:publicKeyConfig:UpdatePublicKey'

-> Text
$sel:id:UpdatePublicKey'

-> UpdatePublicKey

Create a value of UpdatePublicKey with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:ifMatch:UpdatePublicKey', updatePublicKey_ifMatch - The value of the ETag header that you received when retrieving the public key to update. For example: E2QWRUHAPOMQZL.

$sel:publicKeyConfig:UpdatePublicKey', updatePublicKey_publicKeyConfig - A public key configuration.

$sel:id:UpdatePublicKey', updatePublicKey_id - The identifier of the public key that you are updating.

Request Lenses

updatePublicKey_ifMatch :: Lens' UpdatePublicKey (Maybe Text)

The value of the ETag header that you received when retrieving the public key to update. For example: E2QWRUHAPOMQZL.

updatePublicKey_publicKeyConfig :: Lens' UpdatePublicKey PublicKeyConfig

A public key configuration.

updatePublicKey_id :: Lens' UpdatePublicKey Text

The identifier of the public key that you are updating.

Destructuring the Response

data UpdatePublicKeyResponse

See: newUpdatePublicKeyResponse smart constructor.

Constructors

UpdatePublicKeyResponse'

Fields

eTag :: Maybe Text

The identifier of the current version of the public key.

publicKey :: Maybe PublicKey

The public key.

httpStatus :: Int

The response's http status code.

newUpdatePublicKeyResponse

Arguments

:: Int
$sel:httpStatus:UpdatePublicKeyResponse'

-> UpdatePublicKeyResponse

Create a value of UpdatePublicKeyResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:eTag:UpdatePublicKeyResponse', updatePublicKeyResponse_eTag - The identifier of the current version of the public key.

$sel:publicKey:UpdatePublicKeyResponse', updatePublicKeyResponse_publicKey - The public key.

$sel:httpStatus:UpdatePublicKeyResponse', updatePublicKeyResponse_httpStatus - The response's http status code.

Response Lenses

updatePublicKeyResponse_eTag :: Lens' UpdatePublicKeyResponse (Maybe Text)

The identifier of the current version of the public key.

updatePublicKeyResponse_publicKey :: Lens' UpdatePublicKeyResponse (Maybe PublicKey)

The public key.

updatePublicKeyResponse_httpStatus :: Lens' UpdatePublicKeyResponse Int

The response's http status code.