Amazonka.CloudFront.UpdateStreamingDistribution

Description

Update a streaming distribution.

Creating a Request

data UpdateStreamingDistribution

The request to update a streaming distribution.

See: newUpdateStreamingDistribution smart constructor.

Constructors

UpdateStreamingDistribution'

Fields

ifMatch :: Maybe Text

The value of the ETag header that you received when retrieving the streaming distribution's configuration. For example: E2QWRUHAPOMQZL.

streamingDistributionConfig :: StreamingDistributionConfig

The streaming distribution's configuration information.

id :: Text

The streaming distribution's id.

newUpdateStreamingDistribution

Arguments

:: StreamingDistributionConfig
$sel:streamingDistributionConfig:UpdateStreamingDistribution'

-> Text
$sel:id:UpdateStreamingDistribution'

-> UpdateStreamingDistribution

Create a value of UpdateStreamingDistribution with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:ifMatch:UpdateStreamingDistribution', updateStreamingDistribution_ifMatch - The value of the ETag header that you received when retrieving the streaming distribution's configuration. For example: E2QWRUHAPOMQZL.

$sel:streamingDistributionConfig:UpdateStreamingDistribution', updateStreamingDistribution_streamingDistributionConfig - The streaming distribution's configuration information.

$sel:id:UpdateStreamingDistribution', updateStreamingDistribution_id - The streaming distribution's id.

Request Lenses

updateStreamingDistribution_ifMatch :: Lens' UpdateStreamingDistribution (Maybe Text)

The value of the ETag header that you received when retrieving the streaming distribution's configuration. For example: E2QWRUHAPOMQZL.

updateStreamingDistribution_streamingDistributionConfig :: Lens' UpdateStreamingDistribution StreamingDistributionConfig

The streaming distribution's configuration information.

updateStreamingDistribution_id :: Lens' UpdateStreamingDistribution Text

The streaming distribution's id.

Destructuring the Response

data UpdateStreamingDistributionResponse

The returned result of the corresponding request.

See: newUpdateStreamingDistributionResponse smart constructor.

Constructors

UpdateStreamingDistributionResponse'

Fields

eTag :: Maybe Text

The current version of the configuration. For example: E2QWRUHAPOMQZL.

streamingDistribution :: Maybe StreamingDistribution

The streaming distribution's information.

httpStatus :: Int

The response's http status code.

newUpdateStreamingDistributionResponse

Arguments

:: Int
$sel:httpStatus:UpdateStreamingDistributionResponse'

-> UpdateStreamingDistributionResponse

Create a value of UpdateStreamingDistributionResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:eTag:UpdateStreamingDistributionResponse', updateStreamingDistributionResponse_eTag - The current version of the configuration. For example: E2QWRUHAPOMQZL.

$sel:streamingDistribution:UpdateStreamingDistributionResponse', updateStreamingDistributionResponse_streamingDistribution - The streaming distribution's information.

$sel:httpStatus:UpdateStreamingDistributionResponse', updateStreamingDistributionResponse_httpStatus - The response's http status code.

Response Lenses

updateStreamingDistributionResponse_eTag :: Lens' UpdateStreamingDistributionResponse (Maybe Text)

The current version of the configuration. For example: E2QWRUHAPOMQZL.

updateStreamingDistributionResponse_streamingDistribution :: Lens' UpdateStreamingDistributionResponse (Maybe StreamingDistribution)

The streaming distribution's information.

updateStreamingDistributionResponse_httpStatus :: Lens' UpdateStreamingDistributionResponse Int

The response's http status code.