Amazonka.CloudHSM.DescribeLunaClient

Description

This is documentation for AWS CloudHSM Classic. For more information, see AWS CloudHSM Classic FAQs, the AWS CloudHSM Classic User Guide, and the AWS CloudHSM Classic API Reference.

For information about the current version of AWS CloudHSM, see AWS CloudHSM, the AWS CloudHSM User Guide, and the AWS CloudHSM API Reference.

Retrieves information about an HSM client.

Creating a Request

data DescribeLunaClient

See: newDescribeLunaClient smart constructor.

Constructors

DescribeLunaClient'

Fields

clientArn :: Maybe Text

The ARN of the client.

certificateFingerprint :: Maybe Text

The certificate fingerprint.

newDescribeLunaClient :: DescribeLunaClient

Create a value of DescribeLunaClient with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:clientArn:DescribeLunaClient', describeLunaClient_clientArn - The ARN of the client.

$sel:certificateFingerprint:DescribeLunaClient', describeLunaClient_certificateFingerprint - The certificate fingerprint.

Request Lenses

describeLunaClient_clientArn :: Lens' DescribeLunaClient (Maybe Text)

The ARN of the client.

describeLunaClient_certificateFingerprint :: Lens' DescribeLunaClient (Maybe Text)

The certificate fingerprint.

Destructuring the Response

data DescribeLunaClientResponse

See: newDescribeLunaClientResponse smart constructor.

Constructors

DescribeLunaClientResponse'

Fields

clientArn :: Maybe Text

The ARN of the client.

lastModifiedTimestamp :: Maybe Text

The date and time the client was last modified.

certificateFingerprint :: Maybe Text

The certificate fingerprint.

certificate :: Maybe Text

The certificate installed on the HSMs used by this client.

label :: Maybe Text

The label of the client.

httpStatus :: Int

The response's http status code.

newDescribeLunaClientResponse

Arguments

:: Int
$sel:httpStatus:DescribeLunaClientResponse'

-> DescribeLunaClientResponse

Create a value of DescribeLunaClientResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:clientArn:DescribeLunaClient', describeLunaClientResponse_clientArn - The ARN of the client.

$sel:lastModifiedTimestamp:DescribeLunaClientResponse', describeLunaClientResponse_lastModifiedTimestamp - The date and time the client was last modified.

$sel:certificateFingerprint:DescribeLunaClient', describeLunaClientResponse_certificateFingerprint - The certificate fingerprint.

$sel:certificate:DescribeLunaClientResponse', describeLunaClientResponse_certificate - The certificate installed on the HSMs used by this client.

$sel:label:DescribeLunaClientResponse', describeLunaClientResponse_label - The label of the client.

$sel:httpStatus:DescribeLunaClientResponse', describeLunaClientResponse_httpStatus - The response's http status code.

Response Lenses

describeLunaClientResponse_clientArn :: Lens' DescribeLunaClientResponse (Maybe Text)

The ARN of the client.

describeLunaClientResponse_lastModifiedTimestamp :: Lens' DescribeLunaClientResponse (Maybe Text)

The date and time the client was last modified.

describeLunaClientResponse_certificateFingerprint :: Lens' DescribeLunaClientResponse (Maybe Text)

The certificate fingerprint.

describeLunaClientResponse_certificate :: Lens' DescribeLunaClientResponse (Maybe Text)

The certificate installed on the HSMs used by this client.

describeLunaClientResponse_label :: Lens' DescribeLunaClientResponse (Maybe Text)

The label of the client.

describeLunaClientResponse_httpStatus :: Lens' DescribeLunaClientResponse Int

The response's http status code.