Amazonka.CloudHSM.Lens

Description

Operations

DeleteHapg

deleteHapg_hapgArn :: Lens' DeleteHapg Text

The ARN of the high-availability partition group to delete.

deleteHapgResponse_httpStatus :: Lens' DeleteHapgResponse Int

The response's http status code.

deleteHapgResponse_status :: Lens' DeleteHapgResponse Text

The status of the action.

ListHapgs

listHapgs_nextToken :: Lens' ListHapgs (Maybe Text)

The NextToken value from a previous call to ListHapgs. Pass null if this is the first call.

listHapgsResponse_nextToken :: Lens' ListHapgsResponse (Maybe Text)

If not null, more results are available. Pass this value to ListHapgs to retrieve the next set of items.

listHapgsResponse_httpStatus :: Lens' ListHapgsResponse Int

The response's http status code.

listHapgsResponse_hapgList :: Lens' ListHapgsResponse [Text]

The list of high-availability partition groups.

ModifyLunaClient

modifyLunaClient_clientArn :: Lens' ModifyLunaClient Text

The ARN of the client.

modifyLunaClient_certificate :: Lens' ModifyLunaClient Text

The new certificate for the client.

modifyLunaClientResponse_clientArn :: Lens' ModifyLunaClientResponse (Maybe Text)

The ARN of the client.

modifyLunaClientResponse_httpStatus :: Lens' ModifyLunaClientResponse Int

The response's http status code.

ListHsms

listHsms_nextToken :: Lens' ListHsms (Maybe Text)

The NextToken value from a previous call to ListHsms. Pass null if this is the first call.

listHsmsResponse_nextToken :: Lens' ListHsmsResponse (Maybe Text)

If not null, more results are available. Pass this value to ListHsms to retrieve the next set of items.

listHsmsResponse_hsmList :: Lens' ListHsmsResponse (Maybe [Text])

The list of ARNs that identify the HSMs.

listHsmsResponse_httpStatus :: Lens' ListHsmsResponse Int

The response's http status code.

DescribeLunaClient

describeLunaClient_clientArn :: Lens' DescribeLunaClient (Maybe Text)

The ARN of the client.

describeLunaClient_certificateFingerprint :: Lens' DescribeLunaClient (Maybe Text)

The certificate fingerprint.

describeLunaClientResponse_clientArn :: Lens' DescribeLunaClientResponse (Maybe Text)

The ARN of the client.

describeLunaClientResponse_lastModifiedTimestamp :: Lens' DescribeLunaClientResponse (Maybe Text)

The date and time the client was last modified.

describeLunaClientResponse_certificateFingerprint :: Lens' DescribeLunaClientResponse (Maybe Text)

The certificate fingerprint.

describeLunaClientResponse_certificate :: Lens' DescribeLunaClientResponse (Maybe Text)

The certificate installed on the HSMs used by this client.

describeLunaClientResponse_label :: Lens' DescribeLunaClientResponse (Maybe Text)

The label of the client.

describeLunaClientResponse_httpStatus :: Lens' DescribeLunaClientResponse Int

The response's http status code.

ListTagsForResource

listTagsForResource_resourceArn :: Lens' ListTagsForResource Text

The Amazon Resource Name (ARN) of the AWS CloudHSM resource.

listTagsForResourceResponse_httpStatus :: Lens' ListTagsForResourceResponse Int

The response's http status code.

listTagsForResourceResponse_tagList :: Lens' ListTagsForResourceResponse [Tag]

One or more tags.

CreateHapg

createHapg_label :: Lens' CreateHapg Text

The label of the new high-availability partition group.

createHapgResponse_hapgArn :: Lens' CreateHapgResponse (Maybe Text)

The ARN of the high-availability partition group.

createHapgResponse_httpStatus :: Lens' CreateHapgResponse Int

The response's http status code.

CreateHsm

createHsm_clientToken :: Lens' CreateHsm (Maybe Text)

A user-defined token to ensure idempotence. Subsequent calls to this operation with the same token will be ignored.

createHsm_syslogIp :: Lens' CreateHsm (Maybe Text)

The IP address for the syslog monitoring server. The AWS CloudHSM service only supports one syslog monitoring server.

createHsm_externalId :: Lens' CreateHsm (Maybe Text)

The external ID from IamRoleArn, if present.

createHsm_eniIp :: Lens' CreateHsm (Maybe Text)

The IP address to assign to the HSM's ENI.

If an IP address is not specified, an IP address will be randomly chosen from the CIDR range of the subnet.

createHsm_subnetId :: Lens' CreateHsm Text

The identifier of the subnet in your VPC in which to place the HSM.

createHsm_sshKey :: Lens' CreateHsm Text

The SSH public key to install on the HSM.

createHsm_iamRoleArn :: Lens' CreateHsm Text

The ARN of an IAM role to enable the AWS CloudHSM service to allocate an ENI on your behalf.

createHsm_subscriptionType :: Lens' CreateHsm SubscriptionType

Undocumented member.

createHsmResponse_hsmArn :: Lens' CreateHsmResponse (Maybe Text)

The ARN of the HSM.

createHsmResponse_httpStatus :: Lens' CreateHsmResponse Int

The response's http status code.

RemoveTagsFromResource

removeTagsFromResource_resourceArn :: Lens' RemoveTagsFromResource Text

The Amazon Resource Name (ARN) of the AWS CloudHSM resource.

removeTagsFromResource_tagKeyList :: Lens' RemoveTagsFromResource [Text]

The tag key or keys to remove.

Specify only the tag key to remove (not the value). To overwrite the value for an existing tag, use AddTagsToResource.

removeTagsFromResourceResponse_httpStatus :: Lens' RemoveTagsFromResourceResponse Int

The response's http status code.

removeTagsFromResourceResponse_status :: Lens' RemoveTagsFromResourceResponse Text

The status of the operation.

DescribeHapg

describeHapg_hapgArn :: Lens' DescribeHapg Text

The ARN of the high-availability partition group to describe.

describeHapgResponse_state :: Lens' DescribeHapgResponse (Maybe CloudHsmObjectState)

The state of the high-availability partition group.

describeHapgResponse_lastModifiedTimestamp :: Lens' DescribeHapgResponse (Maybe Text)

The date and time the high-availability partition group was last modified.

describeHapgResponse_hsmsPendingRegistration :: Lens' DescribeHapgResponse (Maybe [Text])

describeHapgResponse_hsmsPendingDeletion :: Lens' DescribeHapgResponse (Maybe [Text])

describeHapgResponse_hapgSerial :: Lens' DescribeHapgResponse (Maybe Text)

The serial number of the high-availability partition group.

describeHapgResponse_hsmsLastActionFailed :: Lens' DescribeHapgResponse (Maybe [Text])

describeHapgResponse_partitionSerialList :: Lens' DescribeHapgResponse (Maybe [Text])

The list of partition serial numbers that belong to the high-availability partition group.

describeHapgResponse_hapgArn :: Lens' DescribeHapgResponse (Maybe Text)

The ARN of the high-availability partition group.

describeHapgResponse_label :: Lens' DescribeHapgResponse (Maybe Text)

The label for the high-availability partition group.

describeHapgResponse_httpStatus :: Lens' DescribeHapgResponse Int

The response's http status code.

CreateLunaClient

createLunaClient_label :: Lens' CreateLunaClient (Maybe Text)

The label for the client.

createLunaClient_certificate :: Lens' CreateLunaClient Text

The contents of a Base64-Encoded X.509 v3 certificate to be installed on the HSMs used by this client.

createLunaClientResponse_clientArn :: Lens' CreateLunaClientResponse (Maybe Text)

The ARN of the client.

createLunaClientResponse_httpStatus :: Lens' CreateLunaClientResponse Int

The response's http status code.

ListLunaClients

listLunaClients_nextToken :: Lens' ListLunaClients (Maybe Text)

The NextToken value from a previous call to ListLunaClients. Pass null if this is the first call.

listLunaClientsResponse_nextToken :: Lens' ListLunaClientsResponse (Maybe Text)

If not null, more results are available. Pass this to ListLunaClients to retrieve the next set of items.

listLunaClientsResponse_httpStatus :: Lens' ListLunaClientsResponse Int

The response's http status code.

listLunaClientsResponse_clientList :: Lens' ListLunaClientsResponse [Text]

The list of clients.

AddTagsToResource

addTagsToResource_resourceArn :: Lens' AddTagsToResource Text

The Amazon Resource Name (ARN) of the AWS CloudHSM resource to tag.

addTagsToResource_tagList :: Lens' AddTagsToResource [Tag]

One or more tags.

addTagsToResourceResponse_httpStatus :: Lens' AddTagsToResourceResponse Int

The response's http status code.

addTagsToResourceResponse_status :: Lens' AddTagsToResourceResponse Text

The status of the operation.

GetConfig

getConfig_clientArn :: Lens' GetConfig Text

The ARN of the client.

getConfig_clientVersion :: Lens' GetConfig ClientVersion

The client version.

getConfig_hapgList :: Lens' GetConfig [Text]

A list of ARNs that identify the high-availability partition groups that are associated with the client.

getConfigResponse_configFile :: Lens' GetConfigResponse (Maybe Text)

The chrystoki.conf configuration file.

getConfigResponse_configCred :: Lens' GetConfigResponse (Maybe Text)

The certificate file containing the server.pem files of the HSMs.

getConfigResponse_configType :: Lens' GetConfigResponse (Maybe Text)

The type of credentials.

getConfigResponse_httpStatus :: Lens' GetConfigResponse Int

The response's http status code.

DeleteHsm

deleteHsm_hsmArn :: Lens' DeleteHsm Text

The ARN of the HSM to delete.

deleteHsmResponse_httpStatus :: Lens' DeleteHsmResponse Int

The response's http status code.

deleteHsmResponse_status :: Lens' DeleteHsmResponse Text

The status of the operation.

DescribeHsm

describeHsm_hsmSerialNumber :: Lens' DescribeHsm (Maybe Text)

The serial number of the HSM. Either the HsmArn or the HsmSerialNumber parameter must be specified.

describeHsm_hsmArn :: Lens' DescribeHsm (Maybe Text)

The ARN of the HSM. Either the HsmArn or the SerialNumber parameter must be specified.

describeHsmResponse_status :: Lens' DescribeHsmResponse (Maybe HsmStatus)

The status of the HSM.

describeHsmResponse_iamRoleArn :: Lens' DescribeHsmResponse (Maybe Text)

The ARN of the IAM role assigned to the HSM.

describeHsmResponse_eniId :: Lens' DescribeHsmResponse (Maybe Text)

The identifier of the elastic network interface (ENI) attached to the HSM.

describeHsmResponse_vpcId :: Lens' DescribeHsmResponse (Maybe Text)

The identifier of the VPC that the HSM is in.

describeHsmResponse_sshKeyLastUpdated :: Lens' DescribeHsmResponse (Maybe Text)

The date and time that the SSH key was last updated.

describeHsmResponse_subscriptionEndDate :: Lens' DescribeHsmResponse (Maybe Text)

The subscription end date.

describeHsmResponse_serverCertUri :: Lens' DescribeHsmResponse (Maybe Text)

The URI of the certificate server.

describeHsmResponse_subscriptionType :: Lens' DescribeHsmResponse (Maybe SubscriptionType)

Undocumented member.

describeHsmResponse_sshPublicKey :: Lens' DescribeHsmResponse (Maybe Text)

The public SSH key.

describeHsmResponse_subnetId :: Lens' DescribeHsmResponse (Maybe Text)

The identifier of the subnet that the HSM is in.

describeHsmResponse_statusDetails :: Lens' DescribeHsmResponse (Maybe Text)

Contains additional information about the status of the HSM.

describeHsmResponse_partitions :: Lens' DescribeHsmResponse (Maybe [Text])

The list of partitions on the HSM.

describeHsmResponse_subscriptionStartDate :: Lens' DescribeHsmResponse (Maybe Text)

The subscription start date.

describeHsmResponse_availabilityZone :: Lens' DescribeHsmResponse (Maybe Text)

The Availability Zone that the HSM is in.

describeHsmResponse_serverCertLastUpdated :: Lens' DescribeHsmResponse (Maybe Text)

The date and time that the server certificate was last updated.

describeHsmResponse_softwareVersion :: Lens' DescribeHsmResponse (Maybe Text)

The HSM software version.

describeHsmResponse_vendorName :: Lens' DescribeHsmResponse (Maybe Text)

The name of the HSM vendor.

describeHsmResponse_serialNumber :: Lens' DescribeHsmResponse (Maybe Text)

The serial number of the HSM.

describeHsmResponse_hsmArn :: Lens' DescribeHsmResponse (Maybe Text)

The ARN of the HSM.

describeHsmResponse_eniIp :: Lens' DescribeHsmResponse (Maybe Text)

The IP address assigned to the HSM's ENI.

describeHsmResponse_hsmType :: Lens' DescribeHsmResponse (Maybe Text)

The HSM model type.

describeHsmResponse_httpStatus :: Lens' DescribeHsmResponse Int

The response's http status code.

ModifyHapg

modifyHapg_partitionSerialList :: Lens' ModifyHapg (Maybe [Text])

The list of partition serial numbers to make members of the high-availability partition group.

modifyHapg_label :: Lens' ModifyHapg (Maybe Text)

The new label for the high-availability partition group.

modifyHapg_hapgArn :: Lens' ModifyHapg Text

The ARN of the high-availability partition group to modify.

modifyHapgResponse_hapgArn :: Lens' ModifyHapgResponse (Maybe Text)

The ARN of the high-availability partition group.

modifyHapgResponse_httpStatus :: Lens' ModifyHapgResponse Int

The response's http status code.

DeleteLunaClient

deleteLunaClient_clientArn :: Lens' DeleteLunaClient Text

The ARN of the client to delete.

deleteLunaClientResponse_httpStatus :: Lens' DeleteLunaClientResponse Int

The response's http status code.

deleteLunaClientResponse_status :: Lens' DeleteLunaClientResponse Text

The status of the action.

ModifyHsm

modifyHsm_iamRoleArn :: Lens' ModifyHsm (Maybe Text)

The new IAM role ARN.

modifyHsm_subnetId :: Lens' ModifyHsm (Maybe Text)

The new identifier of the subnet that the HSM is in. The new subnet must be in the same Availability Zone as the current subnet.

modifyHsm_syslogIp :: Lens' ModifyHsm (Maybe Text)

The new IP address for the syslog monitoring server. The AWS CloudHSM service only supports one syslog monitoring server.

modifyHsm_externalId :: Lens' ModifyHsm (Maybe Text)

The new external ID.

modifyHsm_eniIp :: Lens' ModifyHsm (Maybe Text)

The new IP address for the elastic network interface (ENI) attached to the HSM.

If the HSM is moved to a different subnet, and an IP address is not specified, an IP address will be randomly chosen from the CIDR range of the new subnet.

modifyHsm_hsmArn :: Lens' ModifyHsm Text

The ARN of the HSM to modify.

modifyHsmResponse_hsmArn :: Lens' ModifyHsmResponse (Maybe Text)

The ARN of the HSM.

modifyHsmResponse_httpStatus :: Lens' ModifyHsmResponse Int

The response's http status code.

ListAvailableZones

listAvailableZonesResponse_aZList :: Lens' ListAvailableZonesResponse (Maybe [Text])

The list of Availability Zones that have available AWS CloudHSM capacity.

listAvailableZonesResponse_httpStatus :: Lens' ListAvailableZonesResponse Int

The response's http status code.

Types

Tag

tag_key :: Lens' Tag Text

The key of the tag.

tag_value :: Lens' Tag Text

The value of the tag.