Amazonka.CloudHSM.ModifyLunaClient

Description

This is documentation for AWS CloudHSM Classic. For more information, see AWS CloudHSM Classic FAQs, the AWS CloudHSM Classic User Guide, and the AWS CloudHSM Classic API Reference.

For information about the current version of AWS CloudHSM, see AWS CloudHSM, the AWS CloudHSM User Guide, and the AWS CloudHSM API Reference.

Modifies the certificate used by the client.

This action can potentially start a workflow to install the new certificate on the client's HSMs.

Creating a Request

data ModifyLunaClient

See: newModifyLunaClient smart constructor.

Constructors

ModifyLunaClient'

Fields

clientArn :: Text

The ARN of the client.

certificate :: Text

The new certificate for the client.

newModifyLunaClient

Arguments

:: Text
$sel:clientArn:ModifyLunaClient'

-> Text
$sel:certificate:ModifyLunaClient'

-> ModifyLunaClient

Create a value of ModifyLunaClient with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:clientArn:ModifyLunaClient', modifyLunaClient_clientArn - The ARN of the client.

$sel:certificate:ModifyLunaClient', modifyLunaClient_certificate - The new certificate for the client.

Request Lenses

modifyLunaClient_clientArn :: Lens' ModifyLunaClient Text

The ARN of the client.

modifyLunaClient_certificate :: Lens' ModifyLunaClient Text

The new certificate for the client.

Destructuring the Response

data ModifyLunaClientResponse

See: newModifyLunaClientResponse smart constructor.

Constructors

ModifyLunaClientResponse'

Fields

clientArn :: Maybe Text

The ARN of the client.

httpStatus :: Int

The response's http status code.

newModifyLunaClientResponse

Arguments

:: Int
$sel:httpStatus:ModifyLunaClientResponse'

-> ModifyLunaClientResponse

Create a value of ModifyLunaClientResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:clientArn:ModifyLunaClient', modifyLunaClientResponse_clientArn - The ARN of the client.

$sel:httpStatus:ModifyLunaClientResponse', modifyLunaClientResponse_httpStatus - The response's http status code.

Response Lenses

modifyLunaClientResponse_clientArn :: Lens' ModifyLunaClientResponse (Maybe Text)

The ARN of the client.

modifyLunaClientResponse_httpStatus :: Lens' ModifyLunaClientResponse Int

The response's http status code.