Amazonka.CloudHSMV2.CopyBackupToRegion

Description

Copy an AWS CloudHSM cluster backup to a different region.

Creating a Request

data CopyBackupToRegion

See: newCopyBackupToRegion smart constructor.

Constructors

CopyBackupToRegion'

Fields

tagList :: Maybe [Tag]

Tags to apply to the destination backup during creation. If you specify tags, only these tags will be applied to the destination backup. If you do not specify tags, the service copies tags from the source backup to the destination backup.

destinationRegion :: Text

The AWS region that will contain your copied CloudHSM cluster backup.

backupId :: Text

The ID of the backup that will be copied to the destination region.

newCopyBackupToRegion

Arguments

:: Text
$sel:destinationRegion:CopyBackupToRegion'

-> Text
$sel:backupId:CopyBackupToRegion'

-> CopyBackupToRegion

Create a value of CopyBackupToRegion with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:tagList:CopyBackupToRegion', copyBackupToRegion_tagList - Tags to apply to the destination backup during creation. If you specify tags, only these tags will be applied to the destination backup. If you do not specify tags, the service copies tags from the source backup to the destination backup.

$sel:destinationRegion:CopyBackupToRegion', copyBackupToRegion_destinationRegion - The AWS region that will contain your copied CloudHSM cluster backup.

$sel:backupId:CopyBackupToRegion', copyBackupToRegion_backupId - The ID of the backup that will be copied to the destination region.

Request Lenses

copyBackupToRegion_tagList :: Lens' CopyBackupToRegion (Maybe [Tag])

Tags to apply to the destination backup during creation. If you specify tags, only these tags will be applied to the destination backup. If you do not specify tags, the service copies tags from the source backup to the destination backup.

copyBackupToRegion_destinationRegion :: Lens' CopyBackupToRegion Text

The AWS region that will contain your copied CloudHSM cluster backup.

copyBackupToRegion_backupId :: Lens' CopyBackupToRegion Text

The ID of the backup that will be copied to the destination region.

Destructuring the Response

data CopyBackupToRegionResponse

See: newCopyBackupToRegionResponse smart constructor.

Constructors

CopyBackupToRegionResponse'

Fields

destinationBackup :: Maybe DestinationBackup

Information on the backup that will be copied to the destination region, including CreateTimestamp, SourceBackup, SourceCluster, and Source Region. CreateTimestamp of the destination backup will be the same as that of the source backup.

You will need to use the sourceBackupID returned in this operation to use the DescribeBackups operation on the backup that will be copied to the destination region.

httpStatus :: Int

The response's http status code.

newCopyBackupToRegionResponse

Arguments

:: Int
$sel:httpStatus:CopyBackupToRegionResponse'

-> CopyBackupToRegionResponse

Create a value of CopyBackupToRegionResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:destinationBackup:CopyBackupToRegionResponse', copyBackupToRegionResponse_destinationBackup - Information on the backup that will be copied to the destination region, including CreateTimestamp, SourceBackup, SourceCluster, and Source Region. CreateTimestamp of the destination backup will be the same as that of the source backup.

You will need to use the sourceBackupID returned in this operation to use the DescribeBackups operation on the backup that will be copied to the destination region.

$sel:httpStatus:CopyBackupToRegionResponse', copyBackupToRegionResponse_httpStatus - The response's http status code.

Response Lenses

copyBackupToRegionResponse_destinationBackup :: Lens' CopyBackupToRegionResponse (Maybe DestinationBackup)

Information on the backup that will be copied to the destination region, including CreateTimestamp, SourceBackup, SourceCluster, and Source Region. CreateTimestamp of the destination backup will be the same as that of the source backup.

You will need to use the sourceBackupID returned in this operation to use the DescribeBackups operation on the backup that will be copied to the destination region.

copyBackupToRegionResponse_httpStatus :: Lens' CopyBackupToRegionResponse Int

The response's http status code.