Amazonka.CloudSearch.DeleteIndexField

Description

Removes an IndexField from the search domain. For more information, see Configuring Index Fields in the Amazon CloudSearch Developer Guide.

Creating a Request

data DeleteIndexField

Container for the parameters to the DeleteIndexField operation. Specifies the name of the domain you want to update and the name of the index field you want to delete.

See: newDeleteIndexField smart constructor.

Constructors

DeleteIndexField'

Fields

domainName :: Text

indexFieldName :: Text

The name of the index field your want to remove from the domain's indexing options.

newDeleteIndexField

Arguments

:: Text
$sel:domainName:DeleteIndexField'

-> Text
$sel:indexFieldName:DeleteIndexField'

-> DeleteIndexField

Create a value of DeleteIndexField with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:domainName:DeleteIndexField', deleteIndexField_domainName - Undocumented member.

$sel:indexFieldName:DeleteIndexField', deleteIndexField_indexFieldName - The name of the index field your want to remove from the domain's indexing options.

Request Lenses

deleteIndexField_domainName :: Lens' DeleteIndexField Text

Undocumented member.

deleteIndexField_indexFieldName :: Lens' DeleteIndexField Text

The name of the index field your want to remove from the domain's indexing options.

Destructuring the Response

data DeleteIndexFieldResponse

The result of a DeleteIndexField request.

See: newDeleteIndexFieldResponse smart constructor.

Constructors

DeleteIndexFieldResponse'

Fields

httpStatus :: Int

The response's http status code.

indexField :: IndexFieldStatus

The status of the index field being deleted.

newDeleteIndexFieldResponse

Arguments

:: Int
$sel:httpStatus:DeleteIndexFieldResponse'

-> IndexFieldStatus
$sel:indexField:DeleteIndexFieldResponse'

-> DeleteIndexFieldResponse

Create a value of DeleteIndexFieldResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:httpStatus:DeleteIndexFieldResponse', deleteIndexFieldResponse_httpStatus - The response's http status code.

$sel:indexField:DeleteIndexFieldResponse', deleteIndexFieldResponse_indexField - The status of the index field being deleted.

Response Lenses

deleteIndexFieldResponse_httpStatus :: Lens' DeleteIndexFieldResponse Int

The response's http status code.

deleteIndexFieldResponse_indexField :: Lens' DeleteIndexFieldResponse IndexFieldStatus

The status of the index field being deleted.