Amazonka.CloudSearch.DescribeExpressions

Description

Gets the expressions configured for the search domain. Can be limited to specific expressions by name. By default, shows all expressions and includes any pending changes to the configuration. Set the Deployed option to true to show the active configuration and exclude pending changes. For more information, see Configuring Expressions in the Amazon CloudSearch Developer Guide.

Creating a Request

data DescribeExpressions

Container for the parameters to the DescribeDomains operation. Specifies the name of the domain you want to describe. To restrict the response to particular expressions, specify the names of the expressions you want to describe. To show the active configuration and exclude any pending changes, set the Deployed option to true.

See: newDescribeExpressions smart constructor.

Constructors

DescribeExpressions'

Fields

deployed :: Maybe Bool

Whether to display the deployed configuration (true) or include any pending changes (false). Defaults to false.

expressionNames :: Maybe [Text]

Limits the DescribeExpressions response to the specified expressions. If not specified, all expressions are shown.

domainName :: Text

The name of the domain you want to describe.

newDescribeExpressions

Arguments

:: Text
$sel:domainName:DescribeExpressions'

-> DescribeExpressions

Create a value of DescribeExpressions with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:deployed:DescribeExpressions', describeExpressions_deployed - Whether to display the deployed configuration (true) or include any pending changes (false). Defaults to false.

$sel:expressionNames:DescribeExpressions', describeExpressions_expressionNames - Limits the DescribeExpressions response to the specified expressions. If not specified, all expressions are shown.

$sel:domainName:DescribeExpressions', describeExpressions_domainName - The name of the domain you want to describe.

Request Lenses

describeExpressions_deployed :: Lens' DescribeExpressions (Maybe Bool)

Whether to display the deployed configuration (true) or include any pending changes (false). Defaults to false.

describeExpressions_expressionNames :: Lens' DescribeExpressions (Maybe [Text])

Limits the DescribeExpressions response to the specified expressions. If not specified, all expressions are shown.

describeExpressions_domainName :: Lens' DescribeExpressions Text

The name of the domain you want to describe.

Destructuring the Response

data DescribeExpressionsResponse

The result of a DescribeExpressions request. Contains the expressions configured for the domain specified in the request.

See: newDescribeExpressionsResponse smart constructor.

Constructors

DescribeExpressionsResponse'

Fields

httpStatus :: Int

The response's http status code.

expressions :: [ExpressionStatus]

The expressions configured for the domain.

newDescribeExpressionsResponse

Arguments

:: Int
$sel:httpStatus:DescribeExpressionsResponse'

-> DescribeExpressionsResponse

Create a value of DescribeExpressionsResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:httpStatus:DescribeExpressionsResponse', describeExpressionsResponse_httpStatus - The response's http status code.

$sel:expressions:DescribeExpressionsResponse', describeExpressionsResponse_expressions - The expressions configured for the domain.

Response Lenses

describeExpressionsResponse_httpStatus :: Lens' DescribeExpressionsResponse Int

The response's http status code.

describeExpressionsResponse_expressions :: Lens' DescribeExpressionsResponse [ExpressionStatus]

The expressions configured for the domain.