Amazonka.CloudSearch.DescribeSuggesters

Description

Gets the suggesters configured for a domain. A suggester enables you to display possible matches before users finish typing their queries. Can be limited to specific suggesters by name. By default, shows all suggesters and includes any pending changes to the configuration. Set the Deployed option to true to show the active configuration and exclude pending changes. For more information, see Getting Search Suggestions in the Amazon CloudSearch Developer Guide.

Creating a Request

data DescribeSuggesters

Container for the parameters to the DescribeSuggester operation. Specifies the name of the domain you want to describe. To restrict the response to particular suggesters, specify the names of the suggesters you want to describe. To show the active configuration and exclude any pending changes, set the Deployed option to true.

See: newDescribeSuggesters smart constructor.

Constructors

DescribeSuggesters'

Fields

deployed :: Maybe Bool

Whether to display the deployed configuration (true) or include any pending changes (false). Defaults to false.

suggesterNames :: Maybe [Text]

The suggesters you want to describe.

domainName :: Text

The name of the domain you want to describe.

newDescribeSuggesters

Arguments

:: Text
$sel:domainName:DescribeSuggesters'

-> DescribeSuggesters

Create a value of DescribeSuggesters with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:deployed:DescribeSuggesters', describeSuggesters_deployed - Whether to display the deployed configuration (true) or include any pending changes (false). Defaults to false.

$sel:suggesterNames:DescribeSuggesters', describeSuggesters_suggesterNames - The suggesters you want to describe.

$sel:domainName:DescribeSuggesters', describeSuggesters_domainName - The name of the domain you want to describe.

Request Lenses

describeSuggesters_deployed :: Lens' DescribeSuggesters (Maybe Bool)

Whether to display the deployed configuration (true) or include any pending changes (false). Defaults to false.

describeSuggesters_suggesterNames :: Lens' DescribeSuggesters (Maybe [Text])

The suggesters you want to describe.

describeSuggesters_domainName :: Lens' DescribeSuggesters Text

The name of the domain you want to describe.

Destructuring the Response

data DescribeSuggestersResponse

The result of a DescribeSuggesters request.

See: newDescribeSuggestersResponse smart constructor.

Constructors

DescribeSuggestersResponse'

Fields

httpStatus :: Int

The response's http status code.

suggesters :: [SuggesterStatus]

The suggesters configured for the domain specified in the request.

newDescribeSuggestersResponse

Arguments

:: Int
$sel:httpStatus:DescribeSuggestersResponse'

-> DescribeSuggestersResponse

Create a value of DescribeSuggestersResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:httpStatus:DescribeSuggestersResponse', describeSuggestersResponse_httpStatus - The response's http status code.

$sel:suggesters:DescribeSuggestersResponse', describeSuggestersResponse_suggesters - The suggesters configured for the domain specified in the request.

Response Lenses

describeSuggestersResponse_httpStatus :: Lens' DescribeSuggestersResponse Int

The response's http status code.

describeSuggestersResponse_suggesters :: Lens' DescribeSuggestersResponse [SuggesterStatus]

The suggesters configured for the domain specified in the request.