Amazonka.CloudSearch.Lens

Description

Operations

DescribeAvailabilityOptions

describeAvailabilityOptions_deployed :: Lens' DescribeAvailabilityOptions (Maybe Bool)

Whether to display the deployed configuration (true) or include any pending changes (false). Defaults to false.

describeAvailabilityOptions_domainName :: Lens' DescribeAvailabilityOptions Text

The name of the domain you want to describe.

describeAvailabilityOptionsResponse_availabilityOptions :: Lens' DescribeAvailabilityOptionsResponse (Maybe AvailabilityOptionsStatus)

The availability options configured for the domain. Indicates whether Multi-AZ is enabled for the domain.

describeAvailabilityOptionsResponse_httpStatus :: Lens' DescribeAvailabilityOptionsResponse Int

The response's http status code.

DescribeExpressions

describeExpressions_deployed :: Lens' DescribeExpressions (Maybe Bool)

Whether to display the deployed configuration (true) or include any pending changes (false). Defaults to false.

describeExpressions_expressionNames :: Lens' DescribeExpressions (Maybe [Text])

Limits the DescribeExpressions response to the specified expressions. If not specified, all expressions are shown.

describeExpressions_domainName :: Lens' DescribeExpressions Text

The name of the domain you want to describe.

describeExpressionsResponse_httpStatus :: Lens' DescribeExpressionsResponse Int

The response's http status code.

describeExpressionsResponse_expressions :: Lens' DescribeExpressionsResponse [ExpressionStatus]

The expressions configured for the domain.

DefineExpression

defineExpression_domainName :: Lens' DefineExpression Text

Undocumented member.

defineExpression_expression :: Lens' DefineExpression Expression

Undocumented member.

defineExpressionResponse_httpStatus :: Lens' DefineExpressionResponse Int

The response's http status code.

defineExpressionResponse_expression :: Lens' DefineExpressionResponse ExpressionStatus

Undocumented member.

DescribeScalingParameters

describeScalingParameters_domainName :: Lens' DescribeScalingParameters Text

Undocumented member.

describeScalingParametersResponse_httpStatus :: Lens' DescribeScalingParametersResponse Int

The response's http status code.

describeScalingParametersResponse_scalingParameters :: Lens' DescribeScalingParametersResponse ScalingParametersStatus

Undocumented member.

DescribeServiceAccessPolicies

describeServiceAccessPolicies_deployed :: Lens' DescribeServiceAccessPolicies (Maybe Bool)

Whether to display the deployed configuration (true) or include any pending changes (false). Defaults to false.

describeServiceAccessPolicies_domainName :: Lens' DescribeServiceAccessPolicies Text

The name of the domain you want to describe.

describeServiceAccessPoliciesResponse_httpStatus :: Lens' DescribeServiceAccessPoliciesResponse Int

The response's http status code.

describeServiceAccessPoliciesResponse_accessPolicies :: Lens' DescribeServiceAccessPoliciesResponse AccessPoliciesStatus

The access rules configured for the domain specified in the request.

DescribeSuggesters

describeSuggesters_deployed :: Lens' DescribeSuggesters (Maybe Bool)

Whether to display the deployed configuration (true) or include any pending changes (false). Defaults to false.

describeSuggesters_suggesterNames :: Lens' DescribeSuggesters (Maybe [Text])

The suggesters you want to describe.

describeSuggesters_domainName :: Lens' DescribeSuggesters Text

The name of the domain you want to describe.

describeSuggestersResponse_httpStatus :: Lens' DescribeSuggestersResponse Int

The response's http status code.

describeSuggestersResponse_suggesters :: Lens' DescribeSuggestersResponse [SuggesterStatus]

The suggesters configured for the domain specified in the request.

UpdateAvailabilityOptions

updateAvailabilityOptions_domainName :: Lens' UpdateAvailabilityOptions Text

Undocumented member.

updateAvailabilityOptions_multiAZ :: Lens' UpdateAvailabilityOptions Bool

You expand an existing search domain to a second Availability Zone by setting the Multi-AZ option to true. Similarly, you can turn off the Multi-AZ option to downgrade the domain to a single Availability Zone by setting the Multi-AZ option to false.

updateAvailabilityOptionsResponse_availabilityOptions :: Lens' UpdateAvailabilityOptionsResponse (Maybe AvailabilityOptionsStatus)

The newly-configured availability options. Indicates whether Multi-AZ is enabled for the domain.

updateAvailabilityOptionsResponse_httpStatus :: Lens' UpdateAvailabilityOptionsResponse Int

The response's http status code.

DeleteExpression

deleteExpression_domainName :: Lens' DeleteExpression Text

Undocumented member.

deleteExpression_expressionName :: Lens' DeleteExpression Text

The name of the Expression to delete.

deleteExpressionResponse_httpStatus :: Lens' DeleteExpressionResponse Int

The response's http status code.

deleteExpressionResponse_expression :: Lens' DeleteExpressionResponse ExpressionStatus

The status of the expression being deleted.

ListDomainNames

listDomainNamesResponse_domainNames :: Lens' ListDomainNamesResponse (Maybe (HashMap Text Text))

The names of the search domains owned by an account.

listDomainNamesResponse_httpStatus :: Lens' ListDomainNamesResponse Int

The response's http status code.

DefineSuggester

defineSuggester_domainName :: Lens' DefineSuggester Text

Undocumented member.

defineSuggester_suggester :: Lens' DefineSuggester Suggester

Undocumented member.

defineSuggesterResponse_httpStatus :: Lens' DefineSuggesterResponse Int

The response's http status code.

defineSuggesterResponse_suggester :: Lens' DefineSuggesterResponse SuggesterStatus

Undocumented member.

DescribeDomains

describeDomains_domainNames :: Lens' DescribeDomains (Maybe [Text])

The names of the domains you want to include in the response.

describeDomainsResponse_httpStatus :: Lens' DescribeDomainsResponse Int

The response's http status code.

describeDomainsResponse_domainStatusList :: Lens' DescribeDomainsResponse [DomainStatus]

Undocumented member.

DeleteAnalysisScheme

deleteAnalysisScheme_domainName :: Lens' DeleteAnalysisScheme Text

Undocumented member.

deleteAnalysisScheme_analysisSchemeName :: Lens' DeleteAnalysisScheme Text

The name of the analysis scheme you want to delete.

deleteAnalysisSchemeResponse_httpStatus :: Lens' DeleteAnalysisSchemeResponse Int

The response's http status code.

deleteAnalysisSchemeResponse_analysisScheme :: Lens' DeleteAnalysisSchemeResponse AnalysisSchemeStatus

The status of the analysis scheme being deleted.

DescribeDomainEndpointOptions

describeDomainEndpointOptions_deployed :: Lens' DescribeDomainEndpointOptions (Maybe Bool)

Whether to retrieve the latest configuration (which might be in a Processing state) or the current, active configuration. Defaults to false.

describeDomainEndpointOptions_domainName :: Lens' DescribeDomainEndpointOptions Text

A string that represents the name of a domain.

describeDomainEndpointOptionsResponse_domainEndpointOptions :: Lens' DescribeDomainEndpointOptionsResponse (Maybe DomainEndpointOptionsStatus)

The status and configuration of a search domain's endpoint options.

describeDomainEndpointOptionsResponse_httpStatus :: Lens' DescribeDomainEndpointOptionsResponse Int

The response's http status code.

DescribeAnalysisSchemes

describeAnalysisSchemes_deployed :: Lens' DescribeAnalysisSchemes (Maybe Bool)

Whether to display the deployed configuration (true) or include any pending changes (false). Defaults to false.

describeAnalysisSchemes_analysisSchemeNames :: Lens' DescribeAnalysisSchemes (Maybe [Text])

The analysis schemes you want to describe.

describeAnalysisSchemes_domainName :: Lens' DescribeAnalysisSchemes Text

The name of the domain you want to describe.

describeAnalysisSchemesResponse_httpStatus :: Lens' DescribeAnalysisSchemesResponse Int

The response's http status code.

describeAnalysisSchemesResponse_analysisSchemes :: Lens' DescribeAnalysisSchemesResponse [AnalysisSchemeStatus]

The analysis scheme descriptions.

CreateDomain

createDomain_domainName :: Lens' CreateDomain Text

A name for the domain you are creating. Allowed characters are a-z (lower-case letters), 0-9, and hyphen (-). Domain names must start with a letter or number and be at least 3 and no more than 28 characters long.

createDomainResponse_domainStatus :: Lens' CreateDomainResponse (Maybe DomainStatus)

Undocumented member.

createDomainResponse_httpStatus :: Lens' CreateDomainResponse Int

The response's http status code.

UpdateDomainEndpointOptions

updateDomainEndpointOptions_domainName :: Lens' UpdateDomainEndpointOptions Text

A string that represents the name of a domain.

updateDomainEndpointOptions_domainEndpointOptions :: Lens' UpdateDomainEndpointOptions DomainEndpointOptions

Whether to require that all requests to the domain arrive over HTTPS. We recommend Policy-Min-TLS-1-2-2019-07 for TLSSecurityPolicy. For compatibility with older clients, the default is Policy-Min-TLS-1-0-2019-07.

updateDomainEndpointOptionsResponse_domainEndpointOptions :: Lens' UpdateDomainEndpointOptionsResponse (Maybe DomainEndpointOptionsStatus)

The newly-configured domain endpoint options.

updateDomainEndpointOptionsResponse_httpStatus :: Lens' UpdateDomainEndpointOptionsResponse Int

The response's http status code.

DescribeIndexFields

describeIndexFields_deployed :: Lens' DescribeIndexFields (Maybe Bool)

Whether to display the deployed configuration (true) or include any pending changes (false). Defaults to false.

describeIndexFields_fieldNames :: Lens' DescribeIndexFields (Maybe [Text])

A list of the index fields you want to describe. If not specified, information is returned for all configured index fields.

describeIndexFields_domainName :: Lens' DescribeIndexFields Text

The name of the domain you want to describe.

describeIndexFieldsResponse_httpStatus :: Lens' DescribeIndexFieldsResponse Int

The response's http status code.

describeIndexFieldsResponse_indexFields :: Lens' DescribeIndexFieldsResponse [IndexFieldStatus]

The index fields configured for the domain.

DeleteSuggester

deleteSuggester_domainName :: Lens' DeleteSuggester Text

Undocumented member.

deleteSuggester_suggesterName :: Lens' DeleteSuggester Text

Specifies the name of the suggester you want to delete.

deleteSuggesterResponse_httpStatus :: Lens' DeleteSuggesterResponse Int

The response's http status code.

deleteSuggesterResponse_suggester :: Lens' DeleteSuggesterResponse SuggesterStatus

The status of the suggester being deleted.

DefineAnalysisScheme

defineAnalysisScheme_domainName :: Lens' DefineAnalysisScheme Text

Undocumented member.

defineAnalysisScheme_analysisScheme :: Lens' DefineAnalysisScheme AnalysisScheme

Undocumented member.

defineAnalysisSchemeResponse_httpStatus :: Lens' DefineAnalysisSchemeResponse Int

The response's http status code.

defineAnalysisSchemeResponse_analysisScheme :: Lens' DefineAnalysisSchemeResponse AnalysisSchemeStatus

Undocumented member.

IndexDocuments

indexDocuments_domainName :: Lens' IndexDocuments Text

Undocumented member.

indexDocumentsResponse_fieldNames :: Lens' IndexDocumentsResponse (Maybe [Text])

The names of the fields that are currently being indexed.

indexDocumentsResponse_httpStatus :: Lens' IndexDocumentsResponse Int

The response's http status code.

DeleteIndexField

deleteIndexField_domainName :: Lens' DeleteIndexField Text

Undocumented member.

deleteIndexField_indexFieldName :: Lens' DeleteIndexField Text

The name of the index field your want to remove from the domain's indexing options.

deleteIndexFieldResponse_httpStatus :: Lens' DeleteIndexFieldResponse Int

The response's http status code.

deleteIndexFieldResponse_indexField :: Lens' DeleteIndexFieldResponse IndexFieldStatus

The status of the index field being deleted.

UpdateServiceAccessPolicies

updateServiceAccessPolicies_domainName :: Lens' UpdateServiceAccessPolicies Text

Undocumented member.

updateServiceAccessPolicies_accessPolicies :: Lens' UpdateServiceAccessPolicies Text

The access rules you want to configure. These rules replace any existing rules.

updateServiceAccessPoliciesResponse_httpStatus :: Lens' UpdateServiceAccessPoliciesResponse Int

The response's http status code.

updateServiceAccessPoliciesResponse_accessPolicies :: Lens' UpdateServiceAccessPoliciesResponse AccessPoliciesStatus

The access rules configured for the domain.

UpdateScalingParameters

updateScalingParameters_domainName :: Lens' UpdateScalingParameters Text

Undocumented member.

updateScalingParameters_scalingParameters :: Lens' UpdateScalingParameters ScalingParameters

Undocumented member.

updateScalingParametersResponse_httpStatus :: Lens' UpdateScalingParametersResponse Int

The response's http status code.

updateScalingParametersResponse_scalingParameters :: Lens' UpdateScalingParametersResponse ScalingParametersStatus

Undocumented member.

BuildSuggesters

buildSuggesters_domainName :: Lens' BuildSuggesters Text

Undocumented member.

buildSuggestersResponse_fieldNames :: Lens' BuildSuggestersResponse (Maybe [Text])

Undocumented member.

buildSuggestersResponse_httpStatus :: Lens' BuildSuggestersResponse Int

The response's http status code.

DeleteDomain

deleteDomain_domainName :: Lens' DeleteDomain Text

The name of the domain you want to permanently delete.

deleteDomainResponse_domainStatus :: Lens' DeleteDomainResponse (Maybe DomainStatus)

Undocumented member.

deleteDomainResponse_httpStatus :: Lens' DeleteDomainResponse Int

The response's http status code.

DefineIndexField

defineIndexField_domainName :: Lens' DefineIndexField Text

Undocumented member.

defineIndexField_indexField :: Lens' DefineIndexField IndexField

The index field and field options you want to configure.

defineIndexFieldResponse_httpStatus :: Lens' DefineIndexFieldResponse Int

The response's http status code.

defineIndexFieldResponse_indexField :: Lens' DefineIndexFieldResponse IndexFieldStatus

Undocumented member.

Types

AccessPoliciesStatus

accessPoliciesStatus_options :: Lens' AccessPoliciesStatus Text

Undocumented member.

accessPoliciesStatus_status :: Lens' AccessPoliciesStatus OptionStatus

Undocumented member.

AnalysisOptions

analysisOptions_algorithmicStemming :: Lens' AnalysisOptions (Maybe AlgorithmicStemming)

The level of algorithmic stemming to perform: none, minimal, light, or full. The available levels vary depending on the language. For more information, see Language Specific Text Processing Settings in the Amazon CloudSearch Developer Guide

analysisOptions_stopwords :: Lens' AnalysisOptions (Maybe Text)

A JSON array of terms to ignore during indexing and searching. For example, ["a", "an", "the", "of"]. The stopwords dictionary must explicitly list each word you want to ignore. Wildcards and regular expressions are not supported.

analysisOptions_japaneseTokenizationDictionary :: Lens' AnalysisOptions (Maybe Text)

A JSON array that contains a collection of terms, tokens, readings and part of speech for Japanese Tokenizaiton. The Japanese tokenization dictionary enables you to override the default tokenization for selected terms. This is only valid for Japanese language fields.

analysisOptions_synonyms :: Lens' AnalysisOptions (Maybe Text)

A JSON object that defines synonym groups and aliases. A synonym group is an array of arrays, where each sub-array is a group of terms where each term in the group is considered a synonym of every other term in the group. The aliases value is an object that contains a collection of string:value pairs where the string specifies a term and the array of values specifies each of the aliases for that term. An alias is considered a synonym of the specified term, but the term is not considered a synonym of the alias. For more information about specifying synonyms, see Synonyms in the Amazon CloudSearch Developer Guide.

analysisOptions_stemmingDictionary :: Lens' AnalysisOptions (Maybe Text)

A JSON object that contains a collection of string:value pairs that each map a term to its stem. For example, {"term1": "stem1", "term2": "stem2", "term3": "stem3"}. The stemming dictionary is applied in addition to any algorithmic stemming. This enables you to override the results of the algorithmic stemming to correct specific cases of overstemming or understemming. The maximum size of a stemming dictionary is 500 KB.

AnalysisScheme

analysisScheme_analysisOptions :: Lens' AnalysisScheme (Maybe AnalysisOptions)

Undocumented member.

analysisScheme_analysisSchemeName :: Lens' AnalysisScheme Text

Undocumented member.

analysisScheme_analysisSchemeLanguage :: Lens' AnalysisScheme AnalysisSchemeLanguage

Undocumented member.

AnalysisSchemeStatus

analysisSchemeStatus_options :: Lens' AnalysisSchemeStatus AnalysisScheme

Undocumented member.

analysisSchemeStatus_status :: Lens' AnalysisSchemeStatus OptionStatus

Undocumented member.

AvailabilityOptionsStatus

availabilityOptionsStatus_options :: Lens' AvailabilityOptionsStatus Bool

The availability options configured for the domain.

availabilityOptionsStatus_status :: Lens' AvailabilityOptionsStatus OptionStatus

Undocumented member.

DateArrayOptions

dateArrayOptions_sourceFields :: Lens' DateArrayOptions (Maybe Text)

A list of source fields to map to the field.

dateArrayOptions_returnEnabled :: Lens' DateArrayOptions (Maybe Bool)

Whether the contents of the field can be returned in the search results.

dateArrayOptions_facetEnabled :: Lens' DateArrayOptions (Maybe Bool)

Whether facet information can be returned for the field.

dateArrayOptions_searchEnabled :: Lens' DateArrayOptions (Maybe Bool)

Whether the contents of the field are searchable.

dateArrayOptions_defaultValue :: Lens' DateArrayOptions (Maybe Text)

A value to use for the field if the field isn't specified for a document.

DateOptions

dateOptions_sourceField :: Lens' DateOptions (Maybe Text)

Undocumented member.

dateOptions_returnEnabled :: Lens' DateOptions (Maybe Bool)

Whether the contents of the field can be returned in the search results.

dateOptions_facetEnabled :: Lens' DateOptions (Maybe Bool)

Whether facet information can be returned for the field.

dateOptions_searchEnabled :: Lens' DateOptions (Maybe Bool)

Whether the contents of the field are searchable.

dateOptions_sortEnabled :: Lens' DateOptions (Maybe Bool)

Whether the field can be used to sort the search results.

dateOptions_defaultValue :: Lens' DateOptions (Maybe Text)

A value to use for the field if the field isn't specified for a document.

DocumentSuggesterOptions

documentSuggesterOptions_sortExpression :: Lens' DocumentSuggesterOptions (Maybe Text)

An expression that computes a score for each suggestion to control how they are sorted. The scores are rounded to the nearest integer, with a floor of 0 and a ceiling of 2^31-1. A document's relevance score is not computed for suggestions, so sort expressions cannot reference the _score value. To sort suggestions using a numeric field or existing expression, simply specify the name of the field or expression. If no expression is configured for the suggester, the suggestions are sorted with the closest matches listed first.

documentSuggesterOptions_fuzzyMatching :: Lens' DocumentSuggesterOptions (Maybe SuggesterFuzzyMatching)

The level of fuzziness allowed when suggesting matches for a string: none, low, or high. With none, the specified string is treated as an exact prefix. With low, suggestions must differ from the specified string by no more than one character. With high, suggestions can differ by up to two characters. The default is none.

documentSuggesterOptions_sourceField :: Lens' DocumentSuggesterOptions Text

The name of the index field you want to use for suggestions.

DomainEndpointOptions

domainEndpointOptions_enforceHTTPS :: Lens' DomainEndpointOptions (Maybe Bool)

Whether the domain is HTTPS only enabled.

domainEndpointOptions_tLSSecurityPolicy :: Lens' DomainEndpointOptions (Maybe TLSSecurityPolicy)

The minimum required TLS version

DomainEndpointOptionsStatus

domainEndpointOptionsStatus_options :: Lens' DomainEndpointOptionsStatus DomainEndpointOptions

The domain endpoint options configured for the domain.

domainEndpointOptionsStatus_status :: Lens' DomainEndpointOptionsStatus OptionStatus

The status of the configured domain endpoint options.

DomainStatus

domainStatus_searchInstanceCount :: Lens' DomainStatus (Maybe Natural)

The number of search instances that are available to process search requests.

domainStatus_searchInstanceType :: Lens' DomainStatus (Maybe Text)

The instance type that is being used to process search requests.

domainStatus_docService :: Lens' DomainStatus (Maybe ServiceEndpoint)

The service endpoint for updating documents in a search domain.

domainStatus_arn :: Lens' DomainStatus (Maybe Text)

Undocumented member.

domainStatus_created :: Lens' DomainStatus (Maybe Bool)

True if the search domain is created. It can take several minutes to initialize a domain when CreateDomain is called. Newly created search domains are returned from DescribeDomains with a false value for Created until domain creation is complete.

domainStatus_searchService :: Lens' DomainStatus (Maybe ServiceEndpoint)

The service endpoint for requesting search results from a search domain.

domainStatus_limits :: Lens' DomainStatus (Maybe Limits)

Undocumented member.

domainStatus_searchPartitionCount :: Lens' DomainStatus (Maybe Natural)

The number of partitions across which the search index is spread.

domainStatus_deleted :: Lens' DomainStatus (Maybe Bool)

True if the search domain has been deleted. The system must clean up resources dedicated to the search domain when DeleteDomain is called. Newly deleted search domains are returned from DescribeDomains with a true value for IsDeleted for several minutes until resource cleanup is complete.

domainStatus_processing :: Lens' DomainStatus (Maybe Bool)

True if processing is being done to activate the current domain configuration.

domainStatus_domainId :: Lens' DomainStatus Text

Undocumented member.

domainStatus_domainName :: Lens' DomainStatus Text

Undocumented member.

domainStatus_requiresIndexDocuments :: Lens' DomainStatus Bool

True if IndexDocuments needs to be called to activate the current domain configuration.

DoubleArrayOptions

doubleArrayOptions_sourceFields :: Lens' DoubleArrayOptions (Maybe Text)

A list of source fields to map to the field.

doubleArrayOptions_returnEnabled :: Lens' DoubleArrayOptions (Maybe Bool)

Whether the contents of the field can be returned in the search results.

doubleArrayOptions_facetEnabled :: Lens' DoubleArrayOptions (Maybe Bool)

Whether facet information can be returned for the field.

doubleArrayOptions_searchEnabled :: Lens' DoubleArrayOptions (Maybe Bool)

Whether the contents of the field are searchable.

doubleArrayOptions_defaultValue :: Lens' DoubleArrayOptions (Maybe Double)

A value to use for the field if the field isn't specified for a document.

DoubleOptions

doubleOptions_sourceField :: Lens' DoubleOptions (Maybe Text)

The name of the source field to map to the field.

doubleOptions_returnEnabled :: Lens' DoubleOptions (Maybe Bool)

Whether the contents of the field can be returned in the search results.

doubleOptions_facetEnabled :: Lens' DoubleOptions (Maybe Bool)

Whether facet information can be returned for the field.

doubleOptions_searchEnabled :: Lens' DoubleOptions (Maybe Bool)

Whether the contents of the field are searchable.

doubleOptions_sortEnabled :: Lens' DoubleOptions (Maybe Bool)

Whether the field can be used to sort the search results.

doubleOptions_defaultValue :: Lens' DoubleOptions (Maybe Double)

A value to use for the field if the field isn't specified for a document. This can be important if you are using the field in an expression and that field is not present in every document.

Expression

expression_expressionName :: Lens' Expression Text

Undocumented member.

expression_expressionValue :: Lens' Expression Text

Undocumented member.

ExpressionStatus

expressionStatus_options :: Lens' ExpressionStatus Expression

The expression that is evaluated for sorting while processing a search request.

expressionStatus_status :: Lens' ExpressionStatus OptionStatus

Undocumented member.

IndexField

indexField_doubleArrayOptions :: Lens' IndexField (Maybe DoubleArrayOptions)

Undocumented member.

indexField_dateOptions :: Lens' IndexField (Maybe DateOptions)

Undocumented member.

indexField_textArrayOptions :: Lens' IndexField (Maybe TextArrayOptions)

Undocumented member.

indexField_doubleOptions :: Lens' IndexField (Maybe DoubleOptions)

Undocumented member.

indexField_textOptions :: Lens' IndexField (Maybe TextOptions)

Undocumented member.

indexField_latLonOptions :: Lens' IndexField (Maybe LatLonOptions)

Undocumented member.

indexField_literalArrayOptions :: Lens' IndexField (Maybe LiteralArrayOptions)

Undocumented member.

indexField_intArrayOptions :: Lens' IndexField (Maybe IntArrayOptions)

Undocumented member.

indexField_dateArrayOptions :: Lens' IndexField (Maybe DateArrayOptions)

Undocumented member.

indexField_intOptions :: Lens' IndexField (Maybe IntOptions)

Undocumented member.

indexField_literalOptions :: Lens' IndexField (Maybe LiteralOptions)

Undocumented member.

indexField_indexFieldName :: Lens' IndexField Text

A string that represents the name of an index field. CloudSearch supports regular index fields as well as dynamic fields. A dynamic field's name defines a pattern that begins or ends with a wildcard. Any document fields that don't map to a regular index field but do match a dynamic field's pattern are configured with the dynamic field's indexing options.

Regular field names begin with a letter and can contain the following characters: a-z (lowercase), 0-9, and _ (underscore). Dynamic field names must begin or end with a wildcard (*). The wildcard can also be the only character in a dynamic field name. Multiple wildcards, and wildcards embedded within a string are not supported.

The name score is reserved and cannot be used as a field name. To reference a document's ID, you can use the name _id.

indexField_indexFieldType :: Lens' IndexField IndexFieldType

Undocumented member.

IndexFieldStatus

indexFieldStatus_options :: Lens' IndexFieldStatus IndexField

Undocumented member.

indexFieldStatus_status :: Lens' IndexFieldStatus OptionStatus

Undocumented member.

IntArrayOptions

intArrayOptions_sourceFields :: Lens' IntArrayOptions (Maybe Text)

A list of source fields to map to the field.

intArrayOptions_returnEnabled :: Lens' IntArrayOptions (Maybe Bool)

Whether the contents of the field can be returned in the search results.

intArrayOptions_facetEnabled :: Lens' IntArrayOptions (Maybe Bool)

Whether facet information can be returned for the field.

intArrayOptions_searchEnabled :: Lens' IntArrayOptions (Maybe Bool)

Whether the contents of the field are searchable.

intArrayOptions_defaultValue :: Lens' IntArrayOptions (Maybe Integer)

A value to use for the field if the field isn't specified for a document.

IntOptions

intOptions_sourceField :: Lens' IntOptions (Maybe Text)

The name of the source field to map to the field.

intOptions_returnEnabled :: Lens' IntOptions (Maybe Bool)

Whether the contents of the field can be returned in the search results.

intOptions_facetEnabled :: Lens' IntOptions (Maybe Bool)

Whether facet information can be returned for the field.

intOptions_searchEnabled :: Lens' IntOptions (Maybe Bool)

Whether the contents of the field are searchable.

intOptions_sortEnabled :: Lens' IntOptions (Maybe Bool)

Whether the field can be used to sort the search results.

intOptions_defaultValue :: Lens' IntOptions (Maybe Integer)

A value to use for the field if the field isn't specified for a document. This can be important if you are using the field in an expression and that field is not present in every document.

LatLonOptions

latLonOptions_sourceField :: Lens' LatLonOptions (Maybe Text)

Undocumented member.

latLonOptions_returnEnabled :: Lens' LatLonOptions (Maybe Bool)

Whether the contents of the field can be returned in the search results.

latLonOptions_facetEnabled :: Lens' LatLonOptions (Maybe Bool)

Whether facet information can be returned for the field.

latLonOptions_searchEnabled :: Lens' LatLonOptions (Maybe Bool)

Whether the contents of the field are searchable.

latLonOptions_sortEnabled :: Lens' LatLonOptions (Maybe Bool)

Whether the field can be used to sort the search results.

latLonOptions_defaultValue :: Lens' LatLonOptions (Maybe Text)

A value to use for the field if the field isn't specified for a document.

Limits

limits_maximumReplicationCount :: Lens' Limits Natural

Undocumented member.

limits_maximumPartitionCount :: Lens' Limits Natural

Undocumented member.

LiteralArrayOptions

literalArrayOptions_sourceFields :: Lens' LiteralArrayOptions (Maybe Text)

A list of source fields to map to the field.

literalArrayOptions_returnEnabled :: Lens' LiteralArrayOptions (Maybe Bool)

Whether the contents of the field can be returned in the search results.

literalArrayOptions_facetEnabled :: Lens' LiteralArrayOptions (Maybe Bool)

Whether facet information can be returned for the field.

literalArrayOptions_searchEnabled :: Lens' LiteralArrayOptions (Maybe Bool)

Whether the contents of the field are searchable.

literalArrayOptions_defaultValue :: Lens' LiteralArrayOptions (Maybe Text)

A value to use for the field if the field isn't specified for a document.

LiteralOptions

literalOptions_sourceField :: Lens' LiteralOptions (Maybe Text)

Undocumented member.

literalOptions_returnEnabled :: Lens' LiteralOptions (Maybe Bool)

Whether the contents of the field can be returned in the search results.

literalOptions_facetEnabled :: Lens' LiteralOptions (Maybe Bool)

Whether facet information can be returned for the field.

literalOptions_searchEnabled :: Lens' LiteralOptions (Maybe Bool)

Whether the contents of the field are searchable.

literalOptions_sortEnabled :: Lens' LiteralOptions (Maybe Bool)

Whether the field can be used to sort the search results.

literalOptions_defaultValue :: Lens' LiteralOptions (Maybe Text)

A value to use for the field if the field isn't specified for a document.

OptionStatus

optionStatus_pendingDeletion :: Lens' OptionStatus (Maybe Bool)

Indicates that the option will be deleted once processing is complete.

optionStatus_updateVersion :: Lens' OptionStatus (Maybe Natural)

A unique integer that indicates when this option was last updated.

optionStatus_creationDate :: Lens' OptionStatus UTCTime

A timestamp for when this option was created.

optionStatus_updateDate :: Lens' OptionStatus UTCTime

A timestamp for when this option was last updated.

optionStatus_state :: Lens' OptionStatus OptionState

The state of processing a change to an option. Possible values:

RequiresIndexDocuments: the option's latest value will not be deployed until IndexDocuments has been called and indexing is complete.

Processing: the option's latest value is in the process of being activated.

Active: the option's latest value is completely deployed.

FailedToValidate: the option value is not compatible with the domain's data and cannot be used to index the data. You must either modify the option value or update or remove the incompatible documents.

ScalingParameters

scalingParameters_desiredInstanceType :: Lens' ScalingParameters (Maybe PartitionInstanceType)

The instance type that you want to preconfigure for your domain. For example, search.m1.small.

scalingParameters_desiredReplicationCount :: Lens' ScalingParameters (Maybe Natural)

The number of replicas you want to preconfigure for each index partition.

scalingParameters_desiredPartitionCount :: Lens' ScalingParameters (Maybe Natural)

The number of partitions you want to preconfigure for your domain. Only valid when you select m2.2xlarge as the desired instance type.

ScalingParametersStatus

scalingParametersStatus_options :: Lens' ScalingParametersStatus ScalingParameters

Undocumented member.

scalingParametersStatus_status :: Lens' ScalingParametersStatus OptionStatus

Undocumented member.

ServiceEndpoint

serviceEndpoint_endpoint :: Lens' ServiceEndpoint (Maybe Text)

Undocumented member.

Suggester

suggester_suggesterName :: Lens' Suggester Text

Undocumented member.

suggester_documentSuggesterOptions :: Lens' Suggester DocumentSuggesterOptions

Undocumented member.

SuggesterStatus

suggesterStatus_options :: Lens' SuggesterStatus Suggester

Undocumented member.

suggesterStatus_status :: Lens' SuggesterStatus OptionStatus

Undocumented member.

TextArrayOptions

textArrayOptions_sourceFields :: Lens' TextArrayOptions (Maybe Text)

A list of source fields to map to the field.

textArrayOptions_returnEnabled :: Lens' TextArrayOptions (Maybe Bool)

Whether the contents of the field can be returned in the search results.

textArrayOptions_analysisScheme :: Lens' TextArrayOptions (Maybe Text)

The name of an analysis scheme for a text-array field.

textArrayOptions_highlightEnabled :: Lens' TextArrayOptions (Maybe Bool)

Whether highlights can be returned for the field.

textArrayOptions_defaultValue :: Lens' TextArrayOptions (Maybe Text)

A value to use for the field if the field isn't specified for a document.

TextOptions

textOptions_sourceField :: Lens' TextOptions (Maybe Text)

Undocumented member.

textOptions_returnEnabled :: Lens' TextOptions (Maybe Bool)

Whether the contents of the field can be returned in the search results.

textOptions_analysisScheme :: Lens' TextOptions (Maybe Text)

The name of an analysis scheme for a text field.

textOptions_highlightEnabled :: Lens' TextOptions (Maybe Bool)

Whether highlights can be returned for the field.

textOptions_sortEnabled :: Lens' TextOptions (Maybe Bool)

Whether the field can be used to sort the search results.

textOptions_defaultValue :: Lens' TextOptions (Maybe Text)

A value to use for the field if the field isn't specified for a document.