Amazonka.CloudTrail.Types.AdvancedEventSelector

Description

Documentation

data AdvancedEventSelector

Advanced event selectors let you create fine-grained selectors for the following CloudTrail event record ﬁelds. They help you control costs by logging only those events that are important to you. For more information about advanced event selectors, see Logging data events for trails in the CloudTrail User Guide.

readOnly

eventSource

eventName

eventCategory

resources.type

resources.ARN

You cannot apply both event selectors and advanced event selectors to a trail.

See: newAdvancedEventSelector smart constructor.

Constructors

AdvancedEventSelector'

Fields

name :: Maybe Text

An optional, descriptive name for an advanced event selector, such as "Log data events for only two S3 buckets".

fieldSelectors :: NonEmpty AdvancedFieldSelector

Contains all selector statements in an advanced event selector.

newAdvancedEventSelector

Arguments

:: NonEmpty AdvancedFieldSelector
$sel:fieldSelectors:AdvancedEventSelector'

-> AdvancedEventSelector

Create a value of AdvancedEventSelector with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:name:AdvancedEventSelector', advancedEventSelector_name - An optional, descriptive name for an advanced event selector, such as "Log data events for only two S3 buckets".

$sel:fieldSelectors:AdvancedEventSelector', advancedEventSelector_fieldSelectors - Contains all selector statements in an advanced event selector.

advancedEventSelector_name :: Lens' AdvancedEventSelector (Maybe Text)

An optional, descriptive name for an advanced event selector, such as "Log data events for only two S3 buckets".

advancedEventSelector_fieldSelectors :: Lens' AdvancedEventSelector (NonEmpty AdvancedFieldSelector)

Contains all selector statements in an advanced event selector.