Amazonka.CloudTrail.Types.Trail

Description

Documentation

data Trail

The settings for a trail.

See: newTrail smart constructor.

Constructors

Trail'

Fields

logFileValidationEnabled :: Maybe Bool

Specifies whether log file validation is enabled.

trailARN :: Maybe Text

Specifies the ARN of the trail. The following is the format of a trail ARN.

arn:aws:cloudtrail:us-east-2:123456789012:trail/MyTrail

s3KeyPrefix :: Maybe Text

Specifies the Amazon S3 key prefix that comes after the name of the bucket you have designated for log file delivery. For more information, see Finding Your CloudTrail Log Files. The maximum length is 200 characters.

hasInsightSelectors :: Maybe Bool

Specifies whether a trail has insight types specified in an InsightSelector list.

snsTopicARN :: Maybe Text

Specifies the ARN of the Amazon SNS topic that CloudTrail uses to send notifications when log files are delivered. The following is the format of a topic ARN.

arn:aws:sns:us-east-2:123456789012:MyTopic

snsTopicName :: Maybe Text

This field is no longer in use. Use SnsTopicARN.

cloudWatchLogsLogGroupArn :: Maybe Text

Specifies an Amazon Resource Name (ARN), a unique identifier that represents the log group to which CloudTrail logs will be delivered.

kmsKeyId :: Maybe Text

Specifies the KMS key ID that encrypts the logs delivered by CloudTrail. The value is a fully specified ARN to a KMS key in the following format.

arn:aws:kms:us-east-2:123456789012:key/12345678-1234-1234-1234-123456789012

homeRegion :: Maybe Text

The region in which the trail was created.

name :: Maybe Text

Name of the trail set by calling CreateTrail. The maximum length is 128 characters.

includeGlobalServiceEvents :: Maybe Bool

Set to True to include Amazon Web Services API calls from Amazon Web Services global services such as IAM. Otherwise, False.

hasCustomEventSelectors :: Maybe Bool

Specifies if the trail has custom event selectors.

isOrganizationTrail :: Maybe Bool

Specifies whether the trail is an organization trail.

cloudWatchLogsRoleArn :: Maybe Text

Specifies the role for the CloudWatch Logs endpoint to assume to write to a user's log group.

s3BucketName :: Maybe Text

Name of the Amazon S3 bucket into which CloudTrail delivers your trail files. See Amazon S3 Bucket Naming Requirements.

isMultiRegionTrail :: Maybe Bool

Specifies whether the trail exists only in one region or exists in all regions.

newTrail :: Trail

Create a value of Trail with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:logFileValidationEnabled:Trail', trail_logFileValidationEnabled - Specifies whether log file validation is enabled.

$sel:trailARN:Trail', trail_trailARN - Specifies the ARN of the trail. The following is the format of a trail ARN.

arn:aws:cloudtrail:us-east-2:123456789012:trail/MyTrail

$sel:s3KeyPrefix:Trail', trail_s3KeyPrefix - Specifies the Amazon S3 key prefix that comes after the name of the bucket you have designated for log file delivery. For more information, see Finding Your CloudTrail Log Files. The maximum length is 200 characters.

$sel:hasInsightSelectors:Trail', trail_hasInsightSelectors - Specifies whether a trail has insight types specified in an InsightSelector list.

$sel:snsTopicARN:Trail', trail_snsTopicARN - Specifies the ARN of the Amazon SNS topic that CloudTrail uses to send notifications when log files are delivered. The following is the format of a topic ARN.

arn:aws:sns:us-east-2:123456789012:MyTopic

$sel:snsTopicName:Trail', trail_snsTopicName - This field is no longer in use. Use SnsTopicARN.

$sel:cloudWatchLogsLogGroupArn:Trail', trail_cloudWatchLogsLogGroupArn - Specifies an Amazon Resource Name (ARN), a unique identifier that represents the log group to which CloudTrail logs will be delivered.

$sel:kmsKeyId:Trail', trail_kmsKeyId - Specifies the KMS key ID that encrypts the logs delivered by CloudTrail. The value is a fully specified ARN to a KMS key in the following format.

arn:aws:kms:us-east-2:123456789012:key/12345678-1234-1234-1234-123456789012

$sel:homeRegion:Trail', trail_homeRegion - The region in which the trail was created.

$sel:name:Trail', trail_name - Name of the trail set by calling CreateTrail. The maximum length is 128 characters.

$sel:includeGlobalServiceEvents:Trail', trail_includeGlobalServiceEvents - Set to True to include Amazon Web Services API calls from Amazon Web Services global services such as IAM. Otherwise, False.

$sel:hasCustomEventSelectors:Trail', trail_hasCustomEventSelectors - Specifies if the trail has custom event selectors.

$sel:isOrganizationTrail:Trail', trail_isOrganizationTrail - Specifies whether the trail is an organization trail.

$sel:cloudWatchLogsRoleArn:Trail', trail_cloudWatchLogsRoleArn - Specifies the role for the CloudWatch Logs endpoint to assume to write to a user's log group.

$sel:s3BucketName:Trail', trail_s3BucketName - Name of the Amazon S3 bucket into which CloudTrail delivers your trail files. See Amazon S3 Bucket Naming Requirements.

$sel:isMultiRegionTrail:Trail', trail_isMultiRegionTrail - Specifies whether the trail exists only in one region or exists in all regions.

trail_logFileValidationEnabled :: Lens' Trail (Maybe Bool)

Specifies whether log file validation is enabled.

trail_trailARN :: Lens' Trail (Maybe Text)

Specifies the ARN of the trail. The following is the format of a trail ARN.

arn:aws:cloudtrail:us-east-2:123456789012:trail/MyTrail

trail_s3KeyPrefix :: Lens' Trail (Maybe Text)

Specifies the Amazon S3 key prefix that comes after the name of the bucket you have designated for log file delivery. For more information, see Finding Your CloudTrail Log Files. The maximum length is 200 characters.

trail_hasInsightSelectors :: Lens' Trail (Maybe Bool)

Specifies whether a trail has insight types specified in an InsightSelector list.

trail_snsTopicARN :: Lens' Trail (Maybe Text)

Specifies the ARN of the Amazon SNS topic that CloudTrail uses to send notifications when log files are delivered. The following is the format of a topic ARN.

arn:aws:sns:us-east-2:123456789012:MyTopic

trail_snsTopicName :: Lens' Trail (Maybe Text)

This field is no longer in use. Use SnsTopicARN.

trail_cloudWatchLogsLogGroupArn :: Lens' Trail (Maybe Text)

Specifies an Amazon Resource Name (ARN), a unique identifier that represents the log group to which CloudTrail logs will be delivered.

trail_kmsKeyId :: Lens' Trail (Maybe Text)

Specifies the KMS key ID that encrypts the logs delivered by CloudTrail. The value is a fully specified ARN to a KMS key in the following format.

arn:aws:kms:us-east-2:123456789012:key/12345678-1234-1234-1234-123456789012

trail_homeRegion :: Lens' Trail (Maybe Text)

The region in which the trail was created.

trail_name :: Lens' Trail (Maybe Text)

Name of the trail set by calling CreateTrail. The maximum length is 128 characters.

trail_includeGlobalServiceEvents :: Lens' Trail (Maybe Bool)

Set to True to include Amazon Web Services API calls from Amazon Web Services global services such as IAM. Otherwise, False.

trail_hasCustomEventSelectors :: Lens' Trail (Maybe Bool)

Specifies if the trail has custom event selectors.

trail_isOrganizationTrail :: Lens' Trail (Maybe Bool)

Specifies whether the trail is an organization trail.

trail_cloudWatchLogsRoleArn :: Lens' Trail (Maybe Text)

Specifies the role for the CloudWatch Logs endpoint to assume to write to a user's log group.

trail_s3BucketName :: Lens' Trail (Maybe Text)

Name of the Amazon S3 bucket into which CloudTrail delivers your trail files. See Amazon S3 Bucket Naming Requirements.

trail_isMultiRegionTrail :: Lens' Trail (Maybe Bool)

Specifies whether the trail exists only in one region or exists in all regions.