Amazonka.CloudWatch.DescribeAnomalyDetectors

Description

Lists the anomaly detection models that you have created in your account. You can list all models in your account or filter the results to only the models that are related to a certain namespace, metric name, or metric dimension.

Creating a Request

data DescribeAnomalyDetectors

See: newDescribeAnomalyDetectors smart constructor.

Constructors

DescribeAnomalyDetectors'

Fields

metricName :: Maybe Text

Limits the results to only the anomaly detection models that are associated with the specified metric name. If there are multiple metrics with this name in different namespaces that have anomaly detection models, they're all returned.

namespace :: Maybe Text

Limits the results to only the anomaly detection models that are associated with the specified namespace.

nextToken :: Maybe Text

Use the token returned by the previous operation to request the next page of results.

dimensions :: Maybe [Dimension]

Limits the results to only the anomaly detection models that are associated with the specified metric dimensions. If there are multiple metrics that have these dimensions and have anomaly detection models associated, they're all returned.

maxResults :: Maybe Natural

The maximum number of results to return in one operation. The maximum value that you can specify is 100.

To retrieve the remaining results, make another call with the returned NextToken value.

newDescribeAnomalyDetectors :: DescribeAnomalyDetectors

Create a value of DescribeAnomalyDetectors with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:metricName:DescribeAnomalyDetectors', describeAnomalyDetectors_metricName - Limits the results to only the anomaly detection models that are associated with the specified metric name. If there are multiple metrics with this name in different namespaces that have anomaly detection models, they're all returned.

$sel:namespace:DescribeAnomalyDetectors', describeAnomalyDetectors_namespace - Limits the results to only the anomaly detection models that are associated with the specified namespace.

$sel:nextToken:DescribeAnomalyDetectors', describeAnomalyDetectors_nextToken - Use the token returned by the previous operation to request the next page of results.

$sel:dimensions:DescribeAnomalyDetectors', describeAnomalyDetectors_dimensions - Limits the results to only the anomaly detection models that are associated with the specified metric dimensions. If there are multiple metrics that have these dimensions and have anomaly detection models associated, they're all returned.

$sel:maxResults:DescribeAnomalyDetectors', describeAnomalyDetectors_maxResults - The maximum number of results to return in one operation. The maximum value that you can specify is 100.

To retrieve the remaining results, make another call with the returned NextToken value.

Request Lenses

describeAnomalyDetectors_metricName :: Lens' DescribeAnomalyDetectors (Maybe Text)

Limits the results to only the anomaly detection models that are associated with the specified metric name. If there are multiple metrics with this name in different namespaces that have anomaly detection models, they're all returned.

describeAnomalyDetectors_namespace :: Lens' DescribeAnomalyDetectors (Maybe Text)

Limits the results to only the anomaly detection models that are associated with the specified namespace.

describeAnomalyDetectors_nextToken :: Lens' DescribeAnomalyDetectors (Maybe Text)

Use the token returned by the previous operation to request the next page of results.

describeAnomalyDetectors_dimensions :: Lens' DescribeAnomalyDetectors (Maybe [Dimension])

Limits the results to only the anomaly detection models that are associated with the specified metric dimensions. If there are multiple metrics that have these dimensions and have anomaly detection models associated, they're all returned.

describeAnomalyDetectors_maxResults :: Lens' DescribeAnomalyDetectors (Maybe Natural)

The maximum number of results to return in one operation. The maximum value that you can specify is 100.

To retrieve the remaining results, make another call with the returned NextToken value.

Destructuring the Response

data DescribeAnomalyDetectorsResponse

See: newDescribeAnomalyDetectorsResponse smart constructor.

Constructors

DescribeAnomalyDetectorsResponse'

Fields

anomalyDetectors :: Maybe [AnomalyDetector]

The list of anomaly detection models returned by the operation.

nextToken :: Maybe Text

A token that you can use in a subsequent operation to retrieve the next set of results.

httpStatus :: Int

The response's http status code.

newDescribeAnomalyDetectorsResponse

Arguments

:: Int
$sel:httpStatus:DescribeAnomalyDetectorsResponse'

-> DescribeAnomalyDetectorsResponse

Create a value of DescribeAnomalyDetectorsResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:anomalyDetectors:DescribeAnomalyDetectorsResponse', describeAnomalyDetectorsResponse_anomalyDetectors - The list of anomaly detection models returned by the operation.

$sel:nextToken:DescribeAnomalyDetectors', describeAnomalyDetectorsResponse_nextToken - A token that you can use in a subsequent operation to retrieve the next set of results.

$sel:httpStatus:DescribeAnomalyDetectorsResponse', describeAnomalyDetectorsResponse_httpStatus - The response's http status code.

Response Lenses

describeAnomalyDetectorsResponse_anomalyDetectors :: Lens' DescribeAnomalyDetectorsResponse (Maybe [AnomalyDetector])

The list of anomaly detection models returned by the operation.

describeAnomalyDetectorsResponse_nextToken :: Lens' DescribeAnomalyDetectorsResponse (Maybe Text)

A token that you can use in a subsequent operation to retrieve the next set of results.

describeAnomalyDetectorsResponse_httpStatus :: Lens' DescribeAnomalyDetectorsResponse Int

The response's http status code.