Amazonka.CloudWatch.EnableInsightRules

Description

Enables the specified Contributor Insights rules. When rules are enabled, they immediately begin analyzing log data.

Creating a Request

data EnableInsightRules

See: newEnableInsightRules smart constructor.

Constructors

EnableInsightRules'

Fields

ruleNames :: [Text]

An array of the rule names to enable. If you need to find out the names of your rules, use DescribeInsightRules.

newEnableInsightRules :: EnableInsightRules

Create a value of EnableInsightRules with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:ruleNames:EnableInsightRules', enableInsightRules_ruleNames - An array of the rule names to enable. If you need to find out the names of your rules, use DescribeInsightRules.

Request Lenses

enableInsightRules_ruleNames :: Lens' EnableInsightRules [Text]

An array of the rule names to enable. If you need to find out the names of your rules, use DescribeInsightRules.

Destructuring the Response

data EnableInsightRulesResponse

See: newEnableInsightRulesResponse smart constructor.

Constructors

EnableInsightRulesResponse'

Fields

failures :: Maybe [PartialFailure]

An array listing the rules that could not be enabled. You cannot disable or enable built-in rules.

httpStatus :: Int

The response's http status code.

newEnableInsightRulesResponse

Arguments

:: Int
$sel:httpStatus:EnableInsightRulesResponse'

-> EnableInsightRulesResponse

Create a value of EnableInsightRulesResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:failures:EnableInsightRulesResponse', enableInsightRulesResponse_failures - An array listing the rules that could not be enabled. You cannot disable or enable built-in rules.

$sel:httpStatus:EnableInsightRulesResponse', enableInsightRulesResponse_httpStatus - The response's http status code.

Response Lenses

enableInsightRulesResponse_failures :: Lens' EnableInsightRulesResponse (Maybe [PartialFailure])

An array listing the rules that could not be enabled. You cannot disable or enable built-in rules.

enableInsightRulesResponse_httpStatus :: Lens' EnableInsightRulesResponse Int

The response's http status code.