Amazonka.CloudWatch.GetInsightRuleReport

Description

This operation returns the time series data collected by a Contributor Insights rule. The data includes the identity and number of contributors to the log group.

You can also optionally return one or more statistics about each data point in the time series. These statistics can include the following:

UniqueContributors -- the number of unique contributors for each data point.

MaxContributorValue -- the value of the top contributor for each data point. The identity of the contributor might change for each data point in the graph.

If this rule aggregates by COUNT, the top contributor for each data point is the contributor with the most occurrences in that period. If the rule aggregates by SUM, the top contributor is the contributor with the highest sum in the log field specified by the rule's Value, during that period.

SampleCount -- the number of data points matched by the rule.

Sum -- the sum of the values from all contributors during the time period represented by that data point.

Minimum -- the minimum value from a single observation during the time period represented by that data point.

Maximum -- the maximum value from a single observation during the time period represented by that data point.

Average -- the average value from all contributors during the time period represented by that data point.

Creating a Request

data GetInsightRuleReport

See: newGetInsightRuleReport smart constructor.

Constructors

GetInsightRuleReport'

Fields

maxContributorCount :: Maybe Int

The maximum number of contributors to include in the report. The range is 1 to 100. If you omit this, the default of 10 is used.

metrics :: Maybe [Text]

Specifies which metrics to use for aggregation of contributor values for the report. You can specify one or more of the following metrics:

UniqueContributors -- the number of unique contributors for each data point.

MaxContributorValue -- the value of the top contributor for each data point. The identity of the contributor might change for each data point in the graph.

If this rule aggregates by COUNT, the top contributor for each data point is the contributor with the most occurrences in that period. If the rule aggregates by SUM, the top contributor is the contributor with the highest sum in the log field specified by the rule's Value, during that period.

SampleCount -- the number of data points matched by the rule.

Sum -- the sum of the values from all contributors during the time period represented by that data point.

Minimum -- the minimum value from a single observation during the time period represented by that data point.

Maximum -- the maximum value from a single observation during the time period represented by that data point.

Average -- the average value from all contributors during the time period represented by that data point.

orderBy :: Maybe Text

Determines what statistic to use to rank the contributors. Valid values are SUM and MAXIMUM.

ruleName :: Text

The name of the rule that you want to see data from.

startTime :: ISO8601

The start time of the data to use in the report. When used in a raw HTTP Query API, it is formatted as yyyy-MM-dd'T'HH:mm:ss. For example, 2019-07-01T23:59:59.

endTime :: ISO8601

The end time of the data to use in the report. When used in a raw HTTP Query API, it is formatted as yyyy-MM-dd'T'HH:mm:ss. For example, 2019-07-01T23:59:59.

period :: Natural

The period, in seconds, to use for the statistics in the InsightRuleMetricDatapoint results.

newGetInsightRuleReport

Arguments

:: Text
$sel:ruleName:GetInsightRuleReport'

-> UTCTime
$sel:startTime:GetInsightRuleReport'

-> UTCTime
$sel:endTime:GetInsightRuleReport'

-> Natural
$sel:period:GetInsightRuleReport'

-> GetInsightRuleReport

Create a value of GetInsightRuleReport with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:maxContributorCount:GetInsightRuleReport', getInsightRuleReport_maxContributorCount - The maximum number of contributors to include in the report. The range is 1 to 100. If you omit this, the default of 10 is used.

$sel:metrics:GetInsightRuleReport', getInsightRuleReport_metrics - Specifies which metrics to use for aggregation of contributor values for the report. You can specify one or more of the following metrics:

UniqueContributors -- the number of unique contributors for each data point.

MaxContributorValue -- the value of the top contributor for each data point. The identity of the contributor might change for each data point in the graph.

If this rule aggregates by COUNT, the top contributor for each data point is the contributor with the most occurrences in that period. If the rule aggregates by SUM, the top contributor is the contributor with the highest sum in the log field specified by the rule's Value, during that period.

SampleCount -- the number of data points matched by the rule.

Sum -- the sum of the values from all contributors during the time period represented by that data point.

Minimum -- the minimum value from a single observation during the time period represented by that data point.

Maximum -- the maximum value from a single observation during the time period represented by that data point.

Average -- the average value from all contributors during the time period represented by that data point.

$sel:orderBy:GetInsightRuleReport', getInsightRuleReport_orderBy - Determines what statistic to use to rank the contributors. Valid values are SUM and MAXIMUM.

$sel:ruleName:GetInsightRuleReport', getInsightRuleReport_ruleName - The name of the rule that you want to see data from.

$sel:startTime:GetInsightRuleReport', getInsightRuleReport_startTime - The start time of the data to use in the report. When used in a raw HTTP Query API, it is formatted as yyyy-MM-dd'T'HH:mm:ss. For example, 2019-07-01T23:59:59.

$sel:endTime:GetInsightRuleReport', getInsightRuleReport_endTime - The end time of the data to use in the report. When used in a raw HTTP Query API, it is formatted as yyyy-MM-dd'T'HH:mm:ss. For example, 2019-07-01T23:59:59.

$sel:period:GetInsightRuleReport', getInsightRuleReport_period - The period, in seconds, to use for the statistics in the InsightRuleMetricDatapoint results.

Request Lenses

getInsightRuleReport_maxContributorCount :: Lens' GetInsightRuleReport (Maybe Int)

The maximum number of contributors to include in the report. The range is 1 to 100. If you omit this, the default of 10 is used.

getInsightRuleReport_metrics :: Lens' GetInsightRuleReport (Maybe [Text])

Specifies which metrics to use for aggregation of contributor values for the report. You can specify one or more of the following metrics:

UniqueContributors -- the number of unique contributors for each data point.

MaxContributorValue -- the value of the top contributor for each data point. The identity of the contributor might change for each data point in the graph.

If this rule aggregates by COUNT, the top contributor for each data point is the contributor with the most occurrences in that period. If the rule aggregates by SUM, the top contributor is the contributor with the highest sum in the log field specified by the rule's Value, during that period.

SampleCount -- the number of data points matched by the rule.

Sum -- the sum of the values from all contributors during the time period represented by that data point.

Minimum -- the minimum value from a single observation during the time period represented by that data point.

Maximum -- the maximum value from a single observation during the time period represented by that data point.

Average -- the average value from all contributors during the time period represented by that data point.

getInsightRuleReport_orderBy :: Lens' GetInsightRuleReport (Maybe Text)

Determines what statistic to use to rank the contributors. Valid values are SUM and MAXIMUM.

getInsightRuleReport_ruleName :: Lens' GetInsightRuleReport Text

The name of the rule that you want to see data from.

getInsightRuleReport_startTime :: Lens' GetInsightRuleReport UTCTime

The start time of the data to use in the report. When used in a raw HTTP Query API, it is formatted as yyyy-MM-dd'T'HH:mm:ss. For example, 2019-07-01T23:59:59.

getInsightRuleReport_endTime :: Lens' GetInsightRuleReport UTCTime

The end time of the data to use in the report. When used in a raw HTTP Query API, it is formatted as yyyy-MM-dd'T'HH:mm:ss. For example, 2019-07-01T23:59:59.

getInsightRuleReport_period :: Lens' GetInsightRuleReport Natural

The period, in seconds, to use for the statistics in the InsightRuleMetricDatapoint results.

Destructuring the Response

data GetInsightRuleReportResponse

See: newGetInsightRuleReportResponse smart constructor.

Constructors

GetInsightRuleReportResponse'

Fields

keyLabels :: Maybe [Text]

An array of the strings used as the keys for this rule. The keys are the dimensions used to classify contributors. If the rule contains more than one key, then each unique combination of values for the keys is counted as a unique contributor.

approximateUniqueCount :: Maybe Integer

An approximate count of the unique contributors found by this rule in this time period.

aggregationStatistic :: Maybe Text

Specifies whether this rule aggregates contributor data by COUNT or SUM.

aggregateValue :: Maybe Double

The sum of the values from all individual contributors that match the rule.

contributors :: Maybe [InsightRuleContributor]

An array of the unique contributors found by this rule in this time period. If the rule contains multiple keys, each combination of values for the keys counts as a unique contributor.

metricDatapoints :: Maybe [InsightRuleMetricDatapoint]

A time series of metric data points that matches the time period in the rule request.

httpStatus :: Int

The response's http status code.

newGetInsightRuleReportResponse

Arguments

:: Int
$sel:httpStatus:GetInsightRuleReportResponse'

-> GetInsightRuleReportResponse

Create a value of GetInsightRuleReportResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:keyLabels:GetInsightRuleReportResponse', getInsightRuleReportResponse_keyLabels - An array of the strings used as the keys for this rule. The keys are the dimensions used to classify contributors. If the rule contains more than one key, then each unique combination of values for the keys is counted as a unique contributor.

$sel:approximateUniqueCount:GetInsightRuleReportResponse', getInsightRuleReportResponse_approximateUniqueCount - An approximate count of the unique contributors found by this rule in this time period.

$sel:aggregationStatistic:GetInsightRuleReportResponse', getInsightRuleReportResponse_aggregationStatistic - Specifies whether this rule aggregates contributor data by COUNT or SUM.

$sel:aggregateValue:GetInsightRuleReportResponse', getInsightRuleReportResponse_aggregateValue - The sum of the values from all individual contributors that match the rule.

$sel:contributors:GetInsightRuleReportResponse', getInsightRuleReportResponse_contributors - An array of the unique contributors found by this rule in this time period. If the rule contains multiple keys, each combination of values for the keys counts as a unique contributor.

$sel:metricDatapoints:GetInsightRuleReportResponse', getInsightRuleReportResponse_metricDatapoints - A time series of metric data points that matches the time period in the rule request.

$sel:httpStatus:GetInsightRuleReportResponse', getInsightRuleReportResponse_httpStatus - The response's http status code.

Response Lenses

getInsightRuleReportResponse_keyLabels :: Lens' GetInsightRuleReportResponse (Maybe [Text])

An array of the strings used as the keys for this rule. The keys are the dimensions used to classify contributors. If the rule contains more than one key, then each unique combination of values for the keys is counted as a unique contributor.

getInsightRuleReportResponse_approximateUniqueCount :: Lens' GetInsightRuleReportResponse (Maybe Integer)

An approximate count of the unique contributors found by this rule in this time period.

getInsightRuleReportResponse_aggregationStatistic :: Lens' GetInsightRuleReportResponse (Maybe Text)

Specifies whether this rule aggregates contributor data by COUNT or SUM.

getInsightRuleReportResponse_aggregateValue :: Lens' GetInsightRuleReportResponse (Maybe Double)

The sum of the values from all individual contributors that match the rule.

getInsightRuleReportResponse_contributors :: Lens' GetInsightRuleReportResponse (Maybe [InsightRuleContributor])

An array of the unique contributors found by this rule in this time period. If the rule contains multiple keys, each combination of values for the keys counts as a unique contributor.

getInsightRuleReportResponse_metricDatapoints :: Lens' GetInsightRuleReportResponse (Maybe [InsightRuleMetricDatapoint])

A time series of metric data points that matches the time period in the rule request.

getInsightRuleReportResponse_httpStatus :: Lens' GetInsightRuleReportResponse Int

The response's http status code.