Amazonka.CloudWatch.GetMetricStream

Description

Returns information about the metric stream that you specify.

Creating a Request

data GetMetricStream

See: newGetMetricStream smart constructor.

Constructors

GetMetricStream'

Fields

name :: Text

The name of the metric stream to retrieve information about.

newGetMetricStream

Arguments

:: Text
$sel:name:GetMetricStream'

-> GetMetricStream

Create a value of GetMetricStream with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:name:GetMetricStream', getMetricStream_name - The name of the metric stream to retrieve information about.

Request Lenses

getMetricStream_name :: Lens' GetMetricStream Text

The name of the metric stream to retrieve information about.

Destructuring the Response

data GetMetricStreamResponse

See: newGetMetricStreamResponse smart constructor.

Constructors

GetMetricStreamResponse'

Fields

includeFilters :: Maybe [MetricStreamFilter]

If this array of metric namespaces is present, then these namespaces are the only metric namespaces that are streamed by this metric stream.

state :: Maybe Text

The state of the metric stream. The possible values are running and stopped.

excludeFilters :: Maybe [MetricStreamFilter]

If this array of metric namespaces is present, then these namespaces are the only metric namespaces that are not streamed by this metric stream. In this case, all other metric namespaces in the account are streamed by this metric stream.

arn :: Maybe Text

The ARN of the metric stream.

firehoseArn :: Maybe Text

The ARN of the Amazon Kinesis Firehose delivery stream that is used by this metric stream.

outputFormat :: Maybe MetricStreamOutputFormat

lastUpdateDate :: Maybe ISO8601

The date of the most recent update to the metric stream's configuration.

name :: Maybe Text

The name of the metric stream.

creationDate :: Maybe ISO8601

The date that the metric stream was created.

roleArn :: Maybe Text

The ARN of the IAM role that is used by this metric stream.

httpStatus :: Int

The response's http status code.

newGetMetricStreamResponse

Arguments

:: Int
$sel:httpStatus:GetMetricStreamResponse'

-> GetMetricStreamResponse

Create a value of GetMetricStreamResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:includeFilters:GetMetricStreamResponse', getMetricStreamResponse_includeFilters - If this array of metric namespaces is present, then these namespaces are the only metric namespaces that are streamed by this metric stream.

$sel:state:GetMetricStreamResponse', getMetricStreamResponse_state - The state of the metric stream. The possible values are running and stopped.

$sel:excludeFilters:GetMetricStreamResponse', getMetricStreamResponse_excludeFilters - If this array of metric namespaces is present, then these namespaces are the only metric namespaces that are not streamed by this metric stream. In this case, all other metric namespaces in the account are streamed by this metric stream.

$sel:arn:GetMetricStreamResponse', getMetricStreamResponse_arn - The ARN of the metric stream.

$sel:firehoseArn:GetMetricStreamResponse', getMetricStreamResponse_firehoseArn - The ARN of the Amazon Kinesis Firehose delivery stream that is used by this metric stream.

$sel:outputFormat:GetMetricStreamResponse', getMetricStreamResponse_outputFormat -

$sel:lastUpdateDate:GetMetricStreamResponse', getMetricStreamResponse_lastUpdateDate - The date of the most recent update to the metric stream's configuration.

$sel:name:GetMetricStream', getMetricStreamResponse_name - The name of the metric stream.

$sel:creationDate:GetMetricStreamResponse', getMetricStreamResponse_creationDate - The date that the metric stream was created.

$sel:roleArn:GetMetricStreamResponse', getMetricStreamResponse_roleArn - The ARN of the IAM role that is used by this metric stream.

$sel:httpStatus:GetMetricStreamResponse', getMetricStreamResponse_httpStatus - The response's http status code.

Response Lenses

getMetricStreamResponse_includeFilters :: Lens' GetMetricStreamResponse (Maybe [MetricStreamFilter])

If this array of metric namespaces is present, then these namespaces are the only metric namespaces that are streamed by this metric stream.

getMetricStreamResponse_state :: Lens' GetMetricStreamResponse (Maybe Text)

The state of the metric stream. The possible values are running and stopped.

getMetricStreamResponse_excludeFilters :: Lens' GetMetricStreamResponse (Maybe [MetricStreamFilter])

If this array of metric namespaces is present, then these namespaces are the only metric namespaces that are not streamed by this metric stream. In this case, all other metric namespaces in the account are streamed by this metric stream.

getMetricStreamResponse_arn :: Lens' GetMetricStreamResponse (Maybe Text)

The ARN of the metric stream.

getMetricStreamResponse_firehoseArn :: Lens' GetMetricStreamResponse (Maybe Text)

The ARN of the Amazon Kinesis Firehose delivery stream that is used by this metric stream.

getMetricStreamResponse_outputFormat :: Lens' GetMetricStreamResponse (Maybe MetricStreamOutputFormat)

getMetricStreamResponse_lastUpdateDate :: Lens' GetMetricStreamResponse (Maybe UTCTime)

The date of the most recent update to the metric stream's configuration.

getMetricStreamResponse_name :: Lens' GetMetricStreamResponse (Maybe Text)

The name of the metric stream.

getMetricStreamResponse_creationDate :: Lens' GetMetricStreamResponse (Maybe UTCTime)

The date that the metric stream was created.

getMetricStreamResponse_roleArn :: Lens' GetMetricStreamResponse (Maybe Text)

The ARN of the IAM role that is used by this metric stream.

getMetricStreamResponse_httpStatus :: Lens' GetMetricStreamResponse Int

The response's http status code.