Amazonka.CloudWatch.PutAnomalyDetector

Description

Creates an anomaly detection model for a CloudWatch metric. You can use the model to display a band of expected normal values when the metric is graphed.

For more information, see CloudWatch Anomaly Detection.

Creating a Request

data PutAnomalyDetector

See: newPutAnomalyDetector smart constructor.

Constructors

PutAnomalyDetector'

Fields

configuration :: Maybe AnomalyDetectorConfiguration

The configuration specifies details about how the anomaly detection model is to be trained, including time ranges to exclude when training and updating the model. You can specify as many as 10 time ranges.

The configuration can also include the time zone to use for the metric.

dimensions :: Maybe [Dimension]

The metric dimensions to create the anomaly detection model for.

namespace :: Text

The namespace of the metric to create the anomaly detection model for.

metricName :: Text

The name of the metric to create the anomaly detection model for.

stat :: Text

The statistic to use for the metric and the anomaly detection model.

newPutAnomalyDetector

Arguments

:: Text
$sel:namespace:PutAnomalyDetector'

-> Text
$sel:metricName:PutAnomalyDetector'

-> Text
$sel:stat:PutAnomalyDetector'

-> PutAnomalyDetector

Create a value of PutAnomalyDetector with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:configuration:PutAnomalyDetector', putAnomalyDetector_configuration - The configuration specifies details about how the anomaly detection model is to be trained, including time ranges to exclude when training and updating the model. You can specify as many as 10 time ranges.

The configuration can also include the time zone to use for the metric.

$sel:dimensions:PutAnomalyDetector', putAnomalyDetector_dimensions - The metric dimensions to create the anomaly detection model for.

$sel:namespace:PutAnomalyDetector', putAnomalyDetector_namespace - The namespace of the metric to create the anomaly detection model for.

$sel:metricName:PutAnomalyDetector', putAnomalyDetector_metricName - The name of the metric to create the anomaly detection model for.

$sel:stat:PutAnomalyDetector', putAnomalyDetector_stat - The statistic to use for the metric and the anomaly detection model.

Request Lenses

putAnomalyDetector_configuration :: Lens' PutAnomalyDetector (Maybe AnomalyDetectorConfiguration)

The configuration specifies details about how the anomaly detection model is to be trained, including time ranges to exclude when training and updating the model. You can specify as many as 10 time ranges.

The configuration can also include the time zone to use for the metric.

putAnomalyDetector_dimensions :: Lens' PutAnomalyDetector (Maybe [Dimension])

The metric dimensions to create the anomaly detection model for.

putAnomalyDetector_namespace :: Lens' PutAnomalyDetector Text

The namespace of the metric to create the anomaly detection model for.

putAnomalyDetector_metricName :: Lens' PutAnomalyDetector Text

The name of the metric to create the anomaly detection model for.

putAnomalyDetector_stat :: Lens' PutAnomalyDetector Text

The statistic to use for the metric and the anomaly detection model.

Destructuring the Response

data PutAnomalyDetectorResponse

See: newPutAnomalyDetectorResponse smart constructor.

Constructors

PutAnomalyDetectorResponse'

Fields

httpStatus :: Int

The response's http status code.

newPutAnomalyDetectorResponse

Arguments

:: Int
$sel:httpStatus:PutAnomalyDetectorResponse'

-> PutAnomalyDetectorResponse

Create a value of PutAnomalyDetectorResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:httpStatus:PutAnomalyDetectorResponse', putAnomalyDetectorResponse_httpStatus - The response's http status code.

Response Lenses

putAnomalyDetectorResponse_httpStatus :: Lens' PutAnomalyDetectorResponse Int

The response's http status code.