Amazonka.CloudWatch.PutDashboard

Description

Creates a dashboard if it does not already exist, or updates an existing dashboard. If you update a dashboard, the entire contents are replaced with what you specify here.

All dashboards in your account are global, not region-specific.

A simple way to create a dashboard using PutDashboard is to copy an existing dashboard. To copy an existing dashboard using the console, you can load the dashboard and then use the View/edit source command in the Actions menu to display the JSON block for that dashboard. Another way to copy a dashboard is to use GetDashboard, and then use the data returned within DashboardBody as the template for the new dashboard when you call PutDashboard.

When you create a dashboard with PutDashboard, a good practice is to add a text widget at the top of the dashboard with a message that the dashboard was created by script and should not be changed in the console. This message could also point console users to the location of the DashboardBody script or the CloudFormation template used to create the dashboard.

Creating a Request

data PutDashboard

See: newPutDashboard smart constructor.

Constructors

PutDashboard'

Fields

dashboardName :: Text

The name of the dashboard. If a dashboard with this name already exists, this call modifies that dashboard, replacing its current contents. Otherwise, a new dashboard is created. The maximum length is 255, and valid characters are A-Z, a-z, 0-9, "-", and "_". This parameter is required.

dashboardBody :: Text

The detailed information about the dashboard in JSON format, including the widgets to include and their location on the dashboard. This parameter is required.

For more information about the syntax, see Dashboard Body Structure and Syntax.

newPutDashboard

Arguments

:: Text
$sel:dashboardName:PutDashboard'

-> Text
$sel:dashboardBody:PutDashboard'

-> PutDashboard

Create a value of PutDashboard with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:dashboardName:PutDashboard', putDashboard_dashboardName - The name of the dashboard. If a dashboard with this name already exists, this call modifies that dashboard, replacing its current contents. Otherwise, a new dashboard is created. The maximum length is 255, and valid characters are A-Z, a-z, 0-9, "-", and "_". This parameter is required.

$sel:dashboardBody:PutDashboard', putDashboard_dashboardBody - The detailed information about the dashboard in JSON format, including the widgets to include and their location on the dashboard. This parameter is required.

For more information about the syntax, see Dashboard Body Structure and Syntax.

Request Lenses

putDashboard_dashboardName :: Lens' PutDashboard Text

The name of the dashboard. If a dashboard with this name already exists, this call modifies that dashboard, replacing its current contents. Otherwise, a new dashboard is created. The maximum length is 255, and valid characters are A-Z, a-z, 0-9, "-", and "_". This parameter is required.

putDashboard_dashboardBody :: Lens' PutDashboard Text

The detailed information about the dashboard in JSON format, including the widgets to include and their location on the dashboard. This parameter is required.

For more information about the syntax, see Dashboard Body Structure and Syntax.

Destructuring the Response

data PutDashboardResponse

See: newPutDashboardResponse smart constructor.

Constructors

PutDashboardResponse'

Fields

dashboardValidationMessages :: Maybe [DashboardValidationMessage]

If the input for PutDashboard was correct and the dashboard was successfully created or modified, this result is empty.

If this result includes only warning messages, then the input was valid enough for the dashboard to be created or modified, but some elements of the dashboard might not render.

If this result includes error messages, the input was not valid and the operation failed.

httpStatus :: Int

The response's http status code.

newPutDashboardResponse

Arguments

:: Int
$sel:httpStatus:PutDashboardResponse'

-> PutDashboardResponse

Create a value of PutDashboardResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:dashboardValidationMessages:PutDashboardResponse', putDashboardResponse_dashboardValidationMessages - If the input for PutDashboard was correct and the dashboard was successfully created or modified, this result is empty.

If this result includes only warning messages, then the input was valid enough for the dashboard to be created or modified, but some elements of the dashboard might not render.

If this result includes error messages, the input was not valid and the operation failed.

$sel:httpStatus:PutDashboardResponse', putDashboardResponse_httpStatus - The response's http status code.

Response Lenses

putDashboardResponse_dashboardValidationMessages :: Lens' PutDashboardResponse (Maybe [DashboardValidationMessage])

If the input for PutDashboard was correct and the dashboard was successfully created or modified, this result is empty.

If this result includes only warning messages, then the input was valid enough for the dashboard to be created or modified, but some elements of the dashboard might not render.

If this result includes error messages, the input was not valid and the operation failed.

putDashboardResponse_httpStatus :: Lens' PutDashboardResponse Int

The response's http status code.