Amazonka.CloudWatch.PutMetricData

Description

Publishes metric data points to Amazon CloudWatch. CloudWatch associates the data points with the specified metric. If the specified metric does not exist, CloudWatch creates the metric. When CloudWatch creates a metric, it can take up to fifteen minutes for the metric to appear in calls to ListMetrics.

You can publish either individual data points in the Value field, or arrays of values and the number of times each value occurred during the period by using the Values and Counts fields in the MetricDatum structure. Using the Values and Counts method enables you to publish up to 150 values per metric with one PutMetricData request, and supports retrieving percentile statistics on this data.

Each PutMetricData request is limited to 40 KB in size for HTTP POST requests. You can send a payload compressed by gzip. Each request is also limited to no more than 20 different metrics.

Although the Value parameter accepts numbers of type Double, CloudWatch rejects values that are either too small or too large. Values must be in the range of -2^360 to 2^360. In addition, special values (for example, NaN, +Infinity, -Infinity) are not supported.

You can use up to 10 dimensions per metric to further clarify what data the metric collects. Each dimension consists of a Name and Value pair. For more information about specifying dimensions, see Publishing Metrics in the Amazon CloudWatch User Guide.

You specify the time stamp to be associated with each data point. You can specify time stamps that are as much as two weeks before the current date, and as much as 2 hours after the current day and time.

Data points with time stamps from 24 hours ago or longer can take at least 48 hours to become available for GetMetricData or GetMetricStatistics from the time they are submitted. Data points with time stamps between 3 and 24 hours ago can take as much as 2 hours to become available for for GetMetricData or GetMetricStatistics.

CloudWatch needs raw data points to calculate percentile statistics. If you publish data using a statistic set instead, you can only retrieve percentile statistics for this data if one of the following conditions is true:

The SampleCount value of the statistic set is 1 and Min, Max, and Sum are all equal.

The Min and Max are equal, and Sum is equal to Min multiplied by SampleCount.

Creating a Request

data PutMetricData

See: newPutMetricData smart constructor.

Constructors

PutMetricData'

Fields

namespace :: Text

The namespace for the metric data.

To avoid conflicts with Amazon Web Services service namespaces, you should not specify a namespace that begins with AWS/

metricData :: [MetricDatum]

The data for the metric. The array can include no more than 20 metrics per call.

newPutMetricData

Arguments

:: Text
$sel:namespace:PutMetricData'

-> PutMetricData

Create a value of PutMetricData with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:namespace:PutMetricData', putMetricData_namespace - The namespace for the metric data.

To avoid conflicts with Amazon Web Services service namespaces, you should not specify a namespace that begins with AWS/

$sel:metricData:PutMetricData', putMetricData_metricData - The data for the metric. The array can include no more than 20 metrics per call.

Request Lenses

putMetricData_namespace :: Lens' PutMetricData Text

The namespace for the metric data.

To avoid conflicts with Amazon Web Services service namespaces, you should not specify a namespace that begins with AWS/

putMetricData_metricData :: Lens' PutMetricData [MetricDatum]

The data for the metric. The array can include no more than 20 metrics per call.

Destructuring the Response

data PutMetricDataResponse

See: newPutMetricDataResponse smart constructor.

Constructors

PutMetricDataResponse'

newPutMetricDataResponse :: PutMetricDataResponse

Create a value of PutMetricDataResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.