Amazonka.CloudWatch.Types.AlarmHistoryItem

Description

Documentation

data AlarmHistoryItem

Represents the history of a specific alarm.

See: newAlarmHistoryItem smart constructor.

Constructors

AlarmHistoryItem'

Fields

alarmName :: Maybe Text

The descriptive name for the alarm.

historyItemType :: Maybe HistoryItemType

The type of alarm history item.

historyData :: Maybe Text

Data about the alarm, in JSON format.

alarmType :: Maybe AlarmType

The type of alarm, either metric alarm or composite alarm.

historySummary :: Maybe Text

A summary of the alarm history, in text format.

timestamp :: Maybe ISO8601

The time stamp for the alarm history item.

newAlarmHistoryItem :: AlarmHistoryItem

Create a value of AlarmHistoryItem with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:alarmName:AlarmHistoryItem', alarmHistoryItem_alarmName - The descriptive name for the alarm.

$sel:historyItemType:AlarmHistoryItem', alarmHistoryItem_historyItemType - The type of alarm history item.

$sel:historyData:AlarmHistoryItem', alarmHistoryItem_historyData - Data about the alarm, in JSON format.

$sel:alarmType:AlarmHistoryItem', alarmHistoryItem_alarmType - The type of alarm, either metric alarm or composite alarm.

$sel:historySummary:AlarmHistoryItem', alarmHistoryItem_historySummary - A summary of the alarm history, in text format.

$sel:timestamp:AlarmHistoryItem', alarmHistoryItem_timestamp - The time stamp for the alarm history item.

alarmHistoryItem_alarmName :: Lens' AlarmHistoryItem (Maybe Text)

The descriptive name for the alarm.

alarmHistoryItem_historyItemType :: Lens' AlarmHistoryItem (Maybe HistoryItemType)

The type of alarm history item.

alarmHistoryItem_historyData :: Lens' AlarmHistoryItem (Maybe Text)

Data about the alarm, in JSON format.

alarmHistoryItem_alarmType :: Lens' AlarmHistoryItem (Maybe AlarmType)

The type of alarm, either metric alarm or composite alarm.

alarmHistoryItem_historySummary :: Lens' AlarmHistoryItem (Maybe Text)

A summary of the alarm history, in text format.

alarmHistoryItem_timestamp :: Lens' AlarmHistoryItem (Maybe UTCTime)

The time stamp for the alarm history item.