Amazonka.CloudWatch.Types.AnomalyDetectorConfiguration

Description

Documentation

data AnomalyDetectorConfiguration

The configuration specifies details about how the anomaly detection model is to be trained, including time ranges to exclude from use for training the model and the time zone to use for the metric.

See: newAnomalyDetectorConfiguration smart constructor.

Constructors

AnomalyDetectorConfiguration'

Fields

metricTimezone :: Maybe Text

The time zone to use for the metric. This is useful to enable the model to automatically account for daylight savings time changes if the metric is sensitive to such time changes.

To specify a time zone, use the name of the time zone as specified in the standard tz database. For more information, see tz database.

excludedTimeRanges :: Maybe [Range]

An array of time ranges to exclude from use when the anomaly detection model is trained. Use this to make sure that events that could cause unusual values for the metric, such as deployments, aren't used when CloudWatch creates the model.

newAnomalyDetectorConfiguration :: AnomalyDetectorConfiguration

Create a value of AnomalyDetectorConfiguration with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:metricTimezone:AnomalyDetectorConfiguration', anomalyDetectorConfiguration_metricTimezone - The time zone to use for the metric. This is useful to enable the model to automatically account for daylight savings time changes if the metric is sensitive to such time changes.

To specify a time zone, use the name of the time zone as specified in the standard tz database. For more information, see tz database.

$sel:excludedTimeRanges:AnomalyDetectorConfiguration', anomalyDetectorConfiguration_excludedTimeRanges - An array of time ranges to exclude from use when the anomaly detection model is trained. Use this to make sure that events that could cause unusual values for the metric, such as deployments, aren't used when CloudWatch creates the model.

anomalyDetectorConfiguration_metricTimezone :: Lens' AnomalyDetectorConfiguration (Maybe Text)

The time zone to use for the metric. This is useful to enable the model to automatically account for daylight savings time changes if the metric is sensitive to such time changes.

To specify a time zone, use the name of the time zone as specified in the standard tz database. For more information, see tz database.

anomalyDetectorConfiguration_excludedTimeRanges :: Lens' AnomalyDetectorConfiguration (Maybe [Range])

An array of time ranges to exclude from use when the anomaly detection model is trained. Use this to make sure that events that could cause unusual values for the metric, such as deployments, aren't used when CloudWatch creates the model.