Amazonka.CloudWatch.Types.Datapoint

Description

Documentation

data Datapoint

Encapsulates the statistical data that CloudWatch computes from metric data.

See: newDatapoint smart constructor.

Constructors

Datapoint'

Fields

sampleCount :: Maybe Double

The number of metric values that contributed to the aggregate value of this data point.

maximum :: Maybe Double

The maximum metric value for the data point.

average :: Maybe Double

The average of the metric values that correspond to the data point.

minimum :: Maybe Double

The minimum metric value for the data point.

extendedStatistics :: Maybe (HashMap Text Double)

The percentile statistic for the data point.

sum :: Maybe Double

The sum of the metric values for the data point.

unit :: Maybe StandardUnit

The standard unit for the data point.

timestamp :: Maybe ISO8601

The time stamp used for the data point.

newDatapoint :: Datapoint

Create a value of Datapoint with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:sampleCount:Datapoint', datapoint_sampleCount - The number of metric values that contributed to the aggregate value of this data point.

$sel:maximum:Datapoint', datapoint_maximum - The maximum metric value for the data point.

$sel:average:Datapoint', datapoint_average - The average of the metric values that correspond to the data point.

$sel:minimum:Datapoint', datapoint_minimum - The minimum metric value for the data point.

$sel:extendedStatistics:Datapoint', datapoint_extendedStatistics - The percentile statistic for the data point.

$sel:sum:Datapoint', datapoint_sum - The sum of the metric values for the data point.

$sel:unit:Datapoint', datapoint_unit - The standard unit for the data point.

$sel:timestamp:Datapoint', datapoint_timestamp - The time stamp used for the data point.

datapoint_sampleCount :: Lens' Datapoint (Maybe Double)

The number of metric values that contributed to the aggregate value of this data point.

datapoint_maximum :: Lens' Datapoint (Maybe Double)

The maximum metric value for the data point.

datapoint_average :: Lens' Datapoint (Maybe Double)

The average of the metric values that correspond to the data point.

datapoint_minimum :: Lens' Datapoint (Maybe Double)

The minimum metric value for the data point.

datapoint_extendedStatistics :: Lens' Datapoint (Maybe (HashMap Text Double))

The percentile statistic for the data point.

datapoint_sum :: Lens' Datapoint (Maybe Double)

The sum of the metric values for the data point.

datapoint_unit :: Lens' Datapoint (Maybe StandardUnit)

The standard unit for the data point.

datapoint_timestamp :: Lens' Datapoint (Maybe UTCTime)

The time stamp used for the data point.