Amazonka.CloudWatch.Types.InsightRule

Description

Documentation

data InsightRule

This structure contains the definition for a Contributor Insights rule.

See: newInsightRule smart constructor.

Constructors

InsightRule'

Fields

name :: Text

The name of the rule.

state :: Text

Indicates whether the rule is enabled or disabled.

schema :: Text

For rules that you create, this is always {"Name": "CloudWatchLogRule", "Version": 1}. For built-in rules, this is {"Name": "ServiceLogRule", "Version": 1}

definition :: Text

The definition of the rule, as a JSON object. The definition contains the keywords used to define contributors, the value to aggregate on if this rule returns a sum instead of a count, and the filters. For details on the valid syntax, see Contributor Insights Rule Syntax.

newInsightRule

Arguments

:: Text
$sel:name:InsightRule'

-> Text
$sel:state:InsightRule'

-> Text
$sel:schema:InsightRule'

-> Text
$sel:definition:InsightRule'

-> InsightRule

Create a value of InsightRule with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:name:InsightRule', insightRule_name - The name of the rule.

$sel:state:InsightRule', insightRule_state - Indicates whether the rule is enabled or disabled.

$sel:schema:InsightRule', insightRule_schema - For rules that you create, this is always {"Name": "CloudWatchLogRule", "Version": 1}. For built-in rules, this is {"Name": "ServiceLogRule", "Version": 1}

$sel:definition:InsightRule', insightRule_definition - The definition of the rule, as a JSON object. The definition contains the keywords used to define contributors, the value to aggregate on if this rule returns a sum instead of a count, and the filters. For details on the valid syntax, see Contributor Insights Rule Syntax.

insightRule_name :: Lens' InsightRule Text

The name of the rule.

insightRule_state :: Lens' InsightRule Text

Indicates whether the rule is enabled or disabled.

insightRule_schema :: Lens' InsightRule Text

For rules that you create, this is always {"Name": "CloudWatchLogRule", "Version": 1}. For built-in rules, this is {"Name": "ServiceLogRule", "Version": 1}

insightRule_definition :: Lens' InsightRule Text

The definition of the rule, as a JSON object. The definition contains the keywords used to define contributors, the value to aggregate on if this rule returns a sum instead of a count, and the filters. For details on the valid syntax, see Contributor Insights Rule Syntax.