Amazonka.CloudWatch.Types.StatisticSet

Description

Documentation

data StatisticSet

Represents a set of statistics that describes a specific metric.

See: newStatisticSet smart constructor.

Constructors

StatisticSet'

Fields

sampleCount :: Double

The number of samples used for the statistic set.

sum :: Double

The sum of values for the sample set.

minimum :: Double

The minimum value of the sample set.

maximum :: Double

The maximum value of the sample set.

newStatisticSet

Arguments

:: Double
$sel:sampleCount:StatisticSet'

-> Double
$sel:sum:StatisticSet'

-> Double
$sel:minimum:StatisticSet'

-> Double
$sel:maximum:StatisticSet'

-> StatisticSet

Create a value of StatisticSet with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:sampleCount:StatisticSet', statisticSet_sampleCount - The number of samples used for the statistic set.

$sel:sum:StatisticSet', statisticSet_sum - The sum of values for the sample set.

$sel:minimum:StatisticSet', statisticSet_minimum - The minimum value of the sample set.

$sel:maximum:StatisticSet', statisticSet_maximum - The maximum value of the sample set.

statisticSet_sampleCount :: Lens' StatisticSet Double

The number of samples used for the statistic set.

statisticSet_sum :: Lens' StatisticSet Double

The sum of values for the sample set.

statisticSet_minimum :: Lens' StatisticSet Double

The minimum value of the sample set.

statisticSet_maximum :: Lens' StatisticSet Double

The maximum value of the sample set.