Amazonka.CloudWatchEvents.DeleteConnection

Description

Deletes a connection.

Creating a Request

data DeleteConnection

See: newDeleteConnection smart constructor.

Constructors

DeleteConnection'

Fields

name :: Text

The name of the connection to delete.

newDeleteConnection

Arguments

:: Text
$sel:name:DeleteConnection'

-> DeleteConnection

Create a value of DeleteConnection with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:name:DeleteConnection', deleteConnection_name - The name of the connection to delete.

Request Lenses

deleteConnection_name :: Lens' DeleteConnection Text

The name of the connection to delete.

Destructuring the Response

data DeleteConnectionResponse

See: newDeleteConnectionResponse smart constructor.

Constructors

DeleteConnectionResponse'

Fields

creationTime :: Maybe POSIX

A time stamp for the time that the connection was created.

lastModifiedTime :: Maybe POSIX

A time stamp for the time that the connection was last modified before it was deleted.

lastAuthorizedTime :: Maybe POSIX

A time stamp for the time that the connection was last authorized before it wa deleted.

connectionArn :: Maybe Text

The ARN of the connection that was deleted.

connectionState :: Maybe ConnectionState

The state of the connection before it was deleted.

httpStatus :: Int

The response's http status code.

newDeleteConnectionResponse

Arguments

:: Int
$sel:httpStatus:DeleteConnectionResponse'

-> DeleteConnectionResponse

Create a value of DeleteConnectionResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:creationTime:DeleteConnectionResponse', deleteConnectionResponse_creationTime - A time stamp for the time that the connection was created.

$sel:lastModifiedTime:DeleteConnectionResponse', deleteConnectionResponse_lastModifiedTime - A time stamp for the time that the connection was last modified before it was deleted.

$sel:lastAuthorizedTime:DeleteConnectionResponse', deleteConnectionResponse_lastAuthorizedTime - A time stamp for the time that the connection was last authorized before it wa deleted.

$sel:connectionArn:DeleteConnectionResponse', deleteConnectionResponse_connectionArn - The ARN of the connection that was deleted.

$sel:connectionState:DeleteConnectionResponse', deleteConnectionResponse_connectionState - The state of the connection before it was deleted.

$sel:httpStatus:DeleteConnectionResponse', deleteConnectionResponse_httpStatus - The response's http status code.

Response Lenses

deleteConnectionResponse_creationTime :: Lens' DeleteConnectionResponse (Maybe UTCTime)

A time stamp for the time that the connection was created.

deleteConnectionResponse_lastModifiedTime :: Lens' DeleteConnectionResponse (Maybe UTCTime)

A time stamp for the time that the connection was last modified before it was deleted.

deleteConnectionResponse_lastAuthorizedTime :: Lens' DeleteConnectionResponse (Maybe UTCTime)

A time stamp for the time that the connection was last authorized before it wa deleted.

deleteConnectionResponse_connectionArn :: Lens' DeleteConnectionResponse (Maybe Text)

The ARN of the connection that was deleted.

deleteConnectionResponse_connectionState :: Lens' DeleteConnectionResponse (Maybe ConnectionState)

The state of the connection before it was deleted.

deleteConnectionResponse_httpStatus :: Lens' DeleteConnectionResponse Int

The response's http status code.