Amazonka.CloudWatchEvents.DescribeEventBus

Description

Displays details about an event bus in your account. This can include the external Amazon Web Services accounts that are permitted to write events to your default event bus, and the associated policy. For custom event buses and partner event buses, it displays the name, ARN, policy, state, and creation time.

To enable your account to receive events from other accounts on its default event bus, use PutPermission.

For more information about partner event buses, see CreateEventBus.

Creating a Request

data DescribeEventBus

See: newDescribeEventBus smart constructor.

Constructors

DescribeEventBus'

Fields

name :: Maybe Text

The name or ARN of the event bus to show details for. If you omit this, the default event bus is displayed.

newDescribeEventBus :: DescribeEventBus

Create a value of DescribeEventBus with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:name:DescribeEventBus', describeEventBus_name - The name or ARN of the event bus to show details for. If you omit this, the default event bus is displayed.

Request Lenses

describeEventBus_name :: Lens' DescribeEventBus (Maybe Text)

The name or ARN of the event bus to show details for. If you omit this, the default event bus is displayed.

Destructuring the Response

data DescribeEventBusResponse

See: newDescribeEventBusResponse smart constructor.

Constructors

DescribeEventBusResponse'

Fields

arn :: Maybe Text

The Amazon Resource Name (ARN) of the account permitted to write events to the current account.

name :: Maybe Text

The name of the event bus. Currently, this is always default.

policy :: Maybe Text

The policy that enables the external account to send events to your account.

httpStatus :: Int

The response's http status code.

newDescribeEventBusResponse

Arguments

:: Int
$sel:httpStatus:DescribeEventBusResponse'

-> DescribeEventBusResponse

Create a value of DescribeEventBusResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:arn:DescribeEventBusResponse', describeEventBusResponse_arn - The Amazon Resource Name (ARN) of the account permitted to write events to the current account.

$sel:name:DescribeEventBus', describeEventBusResponse_name - The name of the event bus. Currently, this is always default.

$sel:policy:DescribeEventBusResponse', describeEventBusResponse_policy - The policy that enables the external account to send events to your account.

$sel:httpStatus:DescribeEventBusResponse', describeEventBusResponse_httpStatus - The response's http status code.

Response Lenses

describeEventBusResponse_arn :: Lens' DescribeEventBusResponse (Maybe Text)

The Amazon Resource Name (ARN) of the account permitted to write events to the current account.

describeEventBusResponse_name :: Lens' DescribeEventBusResponse (Maybe Text)

The name of the event bus. Currently, this is always default.

describeEventBusResponse_policy :: Lens' DescribeEventBusResponse (Maybe Text)

The policy that enables the external account to send events to your account.

describeEventBusResponse_httpStatus :: Lens' DescribeEventBusResponse Int

The response's http status code.