Amazonka.CloudWatchEvents.ListEventSources

Description

You can use this to see all the partner event sources that have been shared with your Amazon Web Services account. For more information about partner event sources, see CreateEventBus.

Creating a Request

data ListEventSources

See: newListEventSources smart constructor.

Constructors

ListEventSources'

Fields

nextToken :: Maybe Text

The token returned by a previous call to retrieve the next set of results.

namePrefix :: Maybe Text

Specifying this limits the results to only those partner event sources with names that start with the specified prefix.

limit :: Maybe Natural

Specifying this limits the number of results returned by this operation. The operation also returns a NextToken which you can use in a subsequent operation to retrieve the next set of results.

newListEventSources :: ListEventSources

Create a value of ListEventSources with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:nextToken:ListEventSources', listEventSources_nextToken - The token returned by a previous call to retrieve the next set of results.

$sel:namePrefix:ListEventSources', listEventSources_namePrefix - Specifying this limits the results to only those partner event sources with names that start with the specified prefix.

$sel:limit:ListEventSources', listEventSources_limit - Specifying this limits the number of results returned by this operation. The operation also returns a NextToken which you can use in a subsequent operation to retrieve the next set of results.

Request Lenses

listEventSources_nextToken :: Lens' ListEventSources (Maybe Text)

The token returned by a previous call to retrieve the next set of results.

listEventSources_namePrefix :: Lens' ListEventSources (Maybe Text)

Specifying this limits the results to only those partner event sources with names that start with the specified prefix.

listEventSources_limit :: Lens' ListEventSources (Maybe Natural)

Specifying this limits the number of results returned by this operation. The operation also returns a NextToken which you can use in a subsequent operation to retrieve the next set of results.

Destructuring the Response

data ListEventSourcesResponse

See: newListEventSourcesResponse smart constructor.

Constructors

ListEventSourcesResponse'

Fields

nextToken :: Maybe Text

A token you can use in a subsequent operation to retrieve the next set of results.

eventSources :: Maybe [EventSource]

The list of event sources.

httpStatus :: Int

The response's http status code.

newListEventSourcesResponse

Arguments

:: Int
$sel:httpStatus:ListEventSourcesResponse'

-> ListEventSourcesResponse

Create a value of ListEventSourcesResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:nextToken:ListEventSources', listEventSourcesResponse_nextToken - A token you can use in a subsequent operation to retrieve the next set of results.

$sel:eventSources:ListEventSourcesResponse', listEventSourcesResponse_eventSources - The list of event sources.

$sel:httpStatus:ListEventSourcesResponse', listEventSourcesResponse_httpStatus - The response's http status code.

Response Lenses

listEventSourcesResponse_nextToken :: Lens' ListEventSourcesResponse (Maybe Text)

A token you can use in a subsequent operation to retrieve the next set of results.

listEventSourcesResponse_eventSources :: Lens' ListEventSourcesResponse (Maybe [EventSource])

The list of event sources.

listEventSourcesResponse_httpStatus :: Lens' ListEventSourcesResponse Int

The response's http status code.