Amazonka.CloudWatchEvents.ListReplays

Description

Lists your replays. You can either list all the replays or you can provide a prefix to match to the replay names. Filter parameters are exclusive.

Creating a Request

data ListReplays

See: newListReplays smart constructor.

Constructors

ListReplays'

Fields

eventSourceArn :: Maybe Text

The ARN of the archive from which the events are replayed.

state :: Maybe ReplayState

The state of the replay.

nextToken :: Maybe Text

The token returned by a previous call to retrieve the next set of results.

namePrefix :: Maybe Text

A name prefix to filter the replays returned. Only replays with name that match the prefix are returned.

limit :: Maybe Natural

The maximum number of replays to retrieve.

newListReplays :: ListReplays

Create a value of ListReplays with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:eventSourceArn:ListReplays', listReplays_eventSourceArn - The ARN of the archive from which the events are replayed.

$sel:state:ListReplays', listReplays_state - The state of the replay.

$sel:nextToken:ListReplays', listReplays_nextToken - The token returned by a previous call to retrieve the next set of results.

$sel:namePrefix:ListReplays', listReplays_namePrefix - A name prefix to filter the replays returned. Only replays with name that match the prefix are returned.

$sel:limit:ListReplays', listReplays_limit - The maximum number of replays to retrieve.

Request Lenses

listReplays_eventSourceArn :: Lens' ListReplays (Maybe Text)

The ARN of the archive from which the events are replayed.

listReplays_state :: Lens' ListReplays (Maybe ReplayState)

The state of the replay.

listReplays_nextToken :: Lens' ListReplays (Maybe Text)

The token returned by a previous call to retrieve the next set of results.

listReplays_namePrefix :: Lens' ListReplays (Maybe Text)

A name prefix to filter the replays returned. Only replays with name that match the prefix are returned.

listReplays_limit :: Lens' ListReplays (Maybe Natural)

The maximum number of replays to retrieve.

Destructuring the Response

data ListReplaysResponse

See: newListReplaysResponse smart constructor.

Constructors

ListReplaysResponse'

Fields

replays :: Maybe [Replay]

An array of Replay objects that contain information about the replay.

nextToken :: Maybe Text

The token returned by a previous call to retrieve the next set of results.

httpStatus :: Int

The response's http status code.

newListReplaysResponse

Arguments

:: Int
$sel:httpStatus:ListReplaysResponse'

-> ListReplaysResponse

Create a value of ListReplaysResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:replays:ListReplaysResponse', listReplaysResponse_replays - An array of Replay objects that contain information about the replay.

$sel:nextToken:ListReplays', listReplaysResponse_nextToken - The token returned by a previous call to retrieve the next set of results.

$sel:httpStatus:ListReplaysResponse', listReplaysResponse_httpStatus - The response's http status code.

Response Lenses

listReplaysResponse_replays :: Lens' ListReplaysResponse (Maybe [Replay])

An array of Replay objects that contain information about the replay.

listReplaysResponse_nextToken :: Lens' ListReplaysResponse (Maybe Text)

The token returned by a previous call to retrieve the next set of results.

listReplaysResponse_httpStatus :: Lens' ListReplaysResponse Int

The response's http status code.