Amazonka.CloudWatchEvents.ListRuleNamesByTarget

Description

Lists the rules for the specified target. You can see which of the rules in Amazon EventBridge can invoke a specific target in your account.

This operation returns paginated results.

Creating a Request

data ListRuleNamesByTarget

See: newListRuleNamesByTarget smart constructor.

Constructors

ListRuleNamesByTarget'

Fields

nextToken :: Maybe Text

The token returned by a previous call to retrieve the next set of results.

eventBusName :: Maybe Text

The name or ARN of the event bus to list rules for. If you omit this, the default event bus is used.

limit :: Maybe Natural

The maximum number of results to return.

targetArn :: Text

The Amazon Resource Name (ARN) of the target resource.

newListRuleNamesByTarget

Arguments

:: Text
$sel:targetArn:ListRuleNamesByTarget'

-> ListRuleNamesByTarget

Create a value of ListRuleNamesByTarget with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:nextToken:ListRuleNamesByTarget', listRuleNamesByTarget_nextToken - The token returned by a previous call to retrieve the next set of results.

$sel:eventBusName:ListRuleNamesByTarget', listRuleNamesByTarget_eventBusName - The name or ARN of the event bus to list rules for. If you omit this, the default event bus is used.

$sel:limit:ListRuleNamesByTarget', listRuleNamesByTarget_limit - The maximum number of results to return.

$sel:targetArn:ListRuleNamesByTarget', listRuleNamesByTarget_targetArn - The Amazon Resource Name (ARN) of the target resource.

Request Lenses

listRuleNamesByTarget_nextToken :: Lens' ListRuleNamesByTarget (Maybe Text)

The token returned by a previous call to retrieve the next set of results.

listRuleNamesByTarget_eventBusName :: Lens' ListRuleNamesByTarget (Maybe Text)

The name or ARN of the event bus to list rules for. If you omit this, the default event bus is used.

listRuleNamesByTarget_limit :: Lens' ListRuleNamesByTarget (Maybe Natural)

The maximum number of results to return.

listRuleNamesByTarget_targetArn :: Lens' ListRuleNamesByTarget Text

The Amazon Resource Name (ARN) of the target resource.

Destructuring the Response

data ListRuleNamesByTargetResponse

See: newListRuleNamesByTargetResponse smart constructor.

Constructors

ListRuleNamesByTargetResponse'

Fields

ruleNames :: Maybe [Text]

The names of the rules that can invoke the given target.

nextToken :: Maybe Text

Indicates whether there are additional results to retrieve. If there are no more results, the value is null.

httpStatus :: Int

The response's http status code.

newListRuleNamesByTargetResponse

Arguments

:: Int
$sel:httpStatus:ListRuleNamesByTargetResponse'

-> ListRuleNamesByTargetResponse

Create a value of ListRuleNamesByTargetResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:ruleNames:ListRuleNamesByTargetResponse', listRuleNamesByTargetResponse_ruleNames - The names of the rules that can invoke the given target.

$sel:nextToken:ListRuleNamesByTarget', listRuleNamesByTargetResponse_nextToken - Indicates whether there are additional results to retrieve. If there are no more results, the value is null.

$sel:httpStatus:ListRuleNamesByTargetResponse', listRuleNamesByTargetResponse_httpStatus - The response's http status code.

Response Lenses

listRuleNamesByTargetResponse_ruleNames :: Lens' ListRuleNamesByTargetResponse (Maybe [Text])

The names of the rules that can invoke the given target.

listRuleNamesByTargetResponse_nextToken :: Lens' ListRuleNamesByTargetResponse (Maybe Text)

Indicates whether there are additional results to retrieve. If there are no more results, the value is null.

listRuleNamesByTargetResponse_httpStatus :: Lens' ListRuleNamesByTargetResponse Int

The response's http status code.