Amazonka.CloudWatchEvents.PutEvents

Description

Sends custom events to Amazon EventBridge so that they can be matched to rules.

Creating a Request

data PutEvents

See: newPutEvents smart constructor.

Constructors

PutEvents'

Fields

entries :: NonEmpty PutEventsRequestEntry

The entry that defines an event in your system. You can specify several parameters for the entry such as the source and type of the event, resources associated with the event, and so on.

newPutEvents

Arguments

:: NonEmpty PutEventsRequestEntry
$sel:entries:PutEvents'

-> PutEvents

Create a value of PutEvents with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:entries:PutEvents', putEvents_entries - The entry that defines an event in your system. You can specify several parameters for the entry such as the source and type of the event, resources associated with the event, and so on.

Request Lenses

putEvents_entries :: Lens' PutEvents (NonEmpty PutEventsRequestEntry)

The entry that defines an event in your system. You can specify several parameters for the entry such as the source and type of the event, resources associated with the event, and so on.

Destructuring the Response

data PutEventsResponse

See: newPutEventsResponse smart constructor.

Constructors

PutEventsResponse'

Fields

failedEntryCount :: Maybe Int

The number of failed entries.

entries :: Maybe [PutEventsResultEntry]

The successfully and unsuccessfully ingested events results. If the ingestion was successful, the entry has the event ID in it. Otherwise, you can use the error code and error message to identify the problem with the entry.

httpStatus :: Int

The response's http status code.

newPutEventsResponse

Arguments

:: Int
$sel:httpStatus:PutEventsResponse'

-> PutEventsResponse

Create a value of PutEventsResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:failedEntryCount:PutEventsResponse', putEventsResponse_failedEntryCount - The number of failed entries.

$sel:entries:PutEvents', putEventsResponse_entries - The successfully and unsuccessfully ingested events results. If the ingestion was successful, the entry has the event ID in it. Otherwise, you can use the error code and error message to identify the problem with the entry.

$sel:httpStatus:PutEventsResponse', putEventsResponse_httpStatus - The response's http status code.

Response Lenses

putEventsResponse_failedEntryCount :: Lens' PutEventsResponse (Maybe Int)

The number of failed entries.

putEventsResponse_entries :: Lens' PutEventsResponse (Maybe [PutEventsResultEntry])

The successfully and unsuccessfully ingested events results. If the ingestion was successful, the entry has the event ID in it. Otherwise, you can use the error code and error message to identify the problem with the entry.

putEventsResponse_httpStatus :: Lens' PutEventsResponse Int

The response's http status code.