Amazonka.CloudWatchEvents.Types.BatchParameters

Description

Documentation

data BatchParameters

The custom parameters to be used when the target is an Batch job.

See: newBatchParameters smart constructor.

Constructors

BatchParameters'

Fields

retryStrategy :: Maybe BatchRetryStrategy

The retry strategy to use for failed jobs, if the target is an Batch job. The retry strategy is the number of times to retry the failed job execution. Valid values are 1–10. When you specify a retry strategy here, it overrides the retry strategy defined in the job definition.

arrayProperties :: Maybe BatchArrayProperties

The array properties for the submitted job, such as the size of the array. The array size can be between 2 and 10,000. If you specify array properties for a job, it becomes an array job. This parameter is used only if the target is an Batch job.

jobDefinition :: Text

The ARN or name of the job definition to use if the event target is an Batch job. This job definition must already exist.

jobName :: Text

The name to use for this execution of the job, if the target is an Batch job.

newBatchParameters

Arguments

:: Text
$sel:jobDefinition:BatchParameters'

-> Text
$sel:jobName:BatchParameters'

-> BatchParameters

Create a value of BatchParameters with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:retryStrategy:BatchParameters', batchParameters_retryStrategy - The retry strategy to use for failed jobs, if the target is an Batch job. The retry strategy is the number of times to retry the failed job execution. Valid values are 1–10. When you specify a retry strategy here, it overrides the retry strategy defined in the job definition.

$sel:arrayProperties:BatchParameters', batchParameters_arrayProperties - The array properties for the submitted job, such as the size of the array. The array size can be between 2 and 10,000. If you specify array properties for a job, it becomes an array job. This parameter is used only if the target is an Batch job.

$sel:jobDefinition:BatchParameters', batchParameters_jobDefinition - The ARN or name of the job definition to use if the event target is an Batch job. This job definition must already exist.

$sel:jobName:BatchParameters', batchParameters_jobName - The name to use for this execution of the job, if the target is an Batch job.

batchParameters_retryStrategy :: Lens' BatchParameters (Maybe BatchRetryStrategy)

The retry strategy to use for failed jobs, if the target is an Batch job. The retry strategy is the number of times to retry the failed job execution. Valid values are 1–10. When you specify a retry strategy here, it overrides the retry strategy defined in the job definition.

batchParameters_arrayProperties :: Lens' BatchParameters (Maybe BatchArrayProperties)

The array properties for the submitted job, such as the size of the array. The array size can be between 2 and 10,000. If you specify array properties for a job, it becomes an array job. This parameter is used only if the target is an Batch job.

batchParameters_jobDefinition :: Lens' BatchParameters Text

The ARN or name of the job definition to use if the event target is an Batch job. This job definition must already exist.

batchParameters_jobName :: Lens' BatchParameters Text

The name to use for this execution of the job, if the target is an Batch job.