Amazonka.CloudWatchEvents.Types.InputTransformer

Description

Documentation

data InputTransformer

Contains the parameters needed for you to provide custom input to a target based on one or more pieces of data extracted from the event.

See: newInputTransformer smart constructor.

Constructors

InputTransformer'

Fields

inputPathsMap :: Maybe (HashMap Text Text)

Map of JSON paths to be extracted from the event. You can then insert these in the template in InputTemplate to produce the output you want to be sent to the target.

InputPathsMap is an array key-value pairs, where each value is a valid JSON path. You can have as many as 100 key-value pairs. You must use JSON dot notation, not bracket notation.

The keys cannot start with "Amazon Web Services."

inputTemplate :: Text

Input template where you specify placeholders that will be filled with the values of the keys from InputPathsMap to customize the data sent to the target. Enclose each InputPathsMaps value in brackets: <value> The InputTemplate must be valid JSON.

If InputTemplate is a JSON object (surrounded by curly braces), the following restrictions apply:

The placeholder cannot be used as an object key.

The following example shows the syntax for using InputPathsMap and InputTemplate.

 "InputTransformer":

{

"InputPathsMap": {"instance": "$.detail.instance","status": "$.detail.status"},

"InputTemplate": "<instance> is in state <status>"

}

To have the InputTemplate include quote marks within a JSON string, escape each quote marks with a slash, as in the following example:

 "InputTransformer":

{

"InputPathsMap": {"instance": "$.detail.instance","status": "$.detail.status"},

"InputTemplate": "<instance> is in state \"<status>\""

}

The InputTemplate can also be valid JSON with varibles in quotes or out, as in the following example:

 "InputTransformer":

{

"InputPathsMap": {"instance": "$.detail.instance","status": "$.detail.status"},

"InputTemplate": '{"myInstance": <instance>,"myStatus": "<instance> is in state \"<status>\""}'

}

newInputTransformer

Arguments

:: Text
$sel:inputTemplate:InputTransformer'

-> InputTransformer

Create a value of InputTransformer with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:inputPathsMap:InputTransformer', inputTransformer_inputPathsMap - Map of JSON paths to be extracted from the event. You can then insert these in the template in InputTemplate to produce the output you want to be sent to the target.

InputPathsMap is an array key-value pairs, where each value is a valid JSON path. You can have as many as 100 key-value pairs. You must use JSON dot notation, not bracket notation.

The keys cannot start with "Amazon Web Services."

$sel:inputTemplate:InputTransformer', inputTransformer_inputTemplate - Input template where you specify placeholders that will be filled with the values of the keys from InputPathsMap to customize the data sent to the target. Enclose each InputPathsMaps value in brackets: <value> The InputTemplate must be valid JSON.

If InputTemplate is a JSON object (surrounded by curly braces), the following restrictions apply:

The placeholder cannot be used as an object key.

The following example shows the syntax for using InputPathsMap and InputTemplate.

 "InputTransformer":

{

"InputPathsMap": {"instance": "$.detail.instance","status": "$.detail.status"},

"InputTemplate": "<instance> is in state <status>"

}

To have the InputTemplate include quote marks within a JSON string, escape each quote marks with a slash, as in the following example:

 "InputTransformer":

{

"InputPathsMap": {"instance": "$.detail.instance","status": "$.detail.status"},

"InputTemplate": "<instance> is in state \"<status>\""

}

The InputTemplate can also be valid JSON with varibles in quotes or out, as in the following example:

 "InputTransformer":

{

"InputPathsMap": {"instance": "$.detail.instance","status": "$.detail.status"},

"InputTemplate": '{"myInstance": <instance>,"myStatus": "<instance> is in state \"<status>\""}'

}

inputTransformer_inputPathsMap :: Lens' InputTransformer (Maybe (HashMap Text Text))

Map of JSON paths to be extracted from the event. You can then insert these in the template in InputTemplate to produce the output you want to be sent to the target.

InputPathsMap is an array key-value pairs, where each value is a valid JSON path. You can have as many as 100 key-value pairs. You must use JSON dot notation, not bracket notation.

The keys cannot start with "Amazon Web Services."

inputTransformer_inputTemplate :: Lens' InputTransformer Text

Input template where you specify placeholders that will be filled with the values of the keys from InputPathsMap to customize the data sent to the target. Enclose each InputPathsMaps value in brackets: <value> The InputTemplate must be valid JSON.

If InputTemplate is a JSON object (surrounded by curly braces), the following restrictions apply:

The placeholder cannot be used as an object key.

The following example shows the syntax for using InputPathsMap and InputTemplate.

 "InputTransformer":

{

"InputPathsMap": {"instance": "$.detail.instance","status": "$.detail.status"},

"InputTemplate": "<instance> is in state <status>"

}

To have the InputTemplate include quote marks within a JSON string, escape each quote marks with a slash, as in the following example:

 "InputTransformer":

{

"InputPathsMap": {"instance": "$.detail.instance","status": "$.detail.status"},

"InputTemplate": "<instance> is in state \"<status>\""

}

The InputTemplate can also be valid JSON with varibles in quotes or out, as in the following example:

 "InputTransformer":

{

"InputPathsMap": {"instance": "$.detail.instance","status": "$.detail.status"},

"InputTemplate": '{"myInstance": <instance>,"myStatus": "<instance> is in state \"<status>\""}'

}