Amazonka.CloudWatchEvents.Types.PutEventsRequestEntry

Description

Documentation

data PutEventsRequestEntry

Represents an event to be submitted.

See: newPutEventsRequestEntry smart constructor.

Constructors

PutEventsRequestEntry'

Fields

time :: Maybe POSIX

The time stamp of the event, per RFC3339. If no time stamp is provided, the time stamp of the PutEvents call is used.

detailType :: Maybe Text

Free-form string used to decide what fields to expect in the event detail.

resources :: Maybe [Text]

Amazon Web Services resources, identified by Amazon Resource Name (ARN), which the event primarily concerns. Any number, including zero, may be present.

eventBusName :: Maybe Text

The name or ARN of the event bus to receive the event. Only the rules that are associated with this event bus are used to match the event. If you omit this, the default event bus is used.

source :: Maybe Text

The source of the event.

traceHeader :: Maybe Text

An X-Ray trade header, which is an http header (X-Amzn-Trace-Id) that contains the trace-id associated with the event.

To learn more about X-Ray trace headers, see Tracing header in the X-Ray Developer Guide.

detail :: Maybe Text

A valid JSON string. There is no other schema imposed. The JSON string may contain fields and nested subobjects.

newPutEventsRequestEntry :: PutEventsRequestEntry

Create a value of PutEventsRequestEntry with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:time:PutEventsRequestEntry', putEventsRequestEntry_time - The time stamp of the event, per RFC3339. If no time stamp is provided, the time stamp of the PutEvents call is used.

$sel:detailType:PutEventsRequestEntry', putEventsRequestEntry_detailType - Free-form string used to decide what fields to expect in the event detail.

$sel:resources:PutEventsRequestEntry', putEventsRequestEntry_resources - Amazon Web Services resources, identified by Amazon Resource Name (ARN), which the event primarily concerns. Any number, including zero, may be present.

$sel:eventBusName:PutEventsRequestEntry', putEventsRequestEntry_eventBusName - The name or ARN of the event bus to receive the event. Only the rules that are associated with this event bus are used to match the event. If you omit this, the default event bus is used.

$sel:source:PutEventsRequestEntry', putEventsRequestEntry_source - The source of the event.

$sel:traceHeader:PutEventsRequestEntry', putEventsRequestEntry_traceHeader - An X-Ray trade header, which is an http header (X-Amzn-Trace-Id) that contains the trace-id associated with the event.

To learn more about X-Ray trace headers, see Tracing header in the X-Ray Developer Guide.

$sel:detail:PutEventsRequestEntry', putEventsRequestEntry_detail - A valid JSON string. There is no other schema imposed. The JSON string may contain fields and nested subobjects.

putEventsRequestEntry_time :: Lens' PutEventsRequestEntry (Maybe UTCTime)

The time stamp of the event, per RFC3339. If no time stamp is provided, the time stamp of the PutEvents call is used.

putEventsRequestEntry_detailType :: Lens' PutEventsRequestEntry (Maybe Text)

Free-form string used to decide what fields to expect in the event detail.

putEventsRequestEntry_resources :: Lens' PutEventsRequestEntry (Maybe [Text])

Amazon Web Services resources, identified by Amazon Resource Name (ARN), which the event primarily concerns. Any number, including zero, may be present.

putEventsRequestEntry_eventBusName :: Lens' PutEventsRequestEntry (Maybe Text)

The name or ARN of the event bus to receive the event. Only the rules that are associated with this event bus are used to match the event. If you omit this, the default event bus is used.

putEventsRequestEntry_source :: Lens' PutEventsRequestEntry (Maybe Text)

The source of the event.

putEventsRequestEntry_traceHeader :: Lens' PutEventsRequestEntry (Maybe Text)

An X-Ray trade header, which is an http header (X-Amzn-Trace-Id) that contains the trace-id associated with the event.

To learn more about X-Ray trace headers, see Tracing header in the X-Ray Developer Guide.

putEventsRequestEntry_detail :: Lens' PutEventsRequestEntry (Maybe Text)

A valid JSON string. There is no other schema imposed. The JSON string may contain fields and nested subobjects.