Amazonka.CloudWatchEvents.UpdateConnection

Description

Updates settings for a connection.

Creating a Request

data UpdateConnection

See: newUpdateConnection smart constructor.

Constructors

UpdateConnection'

Fields

authParameters :: Maybe UpdateConnectionAuthRequestParameters

The authorization parameters to use for the connection.

authorizationType :: Maybe ConnectionAuthorizationType

The type of authorization to use for the connection.

description :: Maybe Text

A description for the connection.

name :: Text

The name of the connection to update.

newUpdateConnection

Arguments

:: Text
$sel:name:UpdateConnection'

-> UpdateConnection

Create a value of UpdateConnection with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:authParameters:UpdateConnection', updateConnection_authParameters - The authorization parameters to use for the connection.

$sel:authorizationType:UpdateConnection', updateConnection_authorizationType - The type of authorization to use for the connection.

$sel:description:UpdateConnection', updateConnection_description - A description for the connection.

$sel:name:UpdateConnection', updateConnection_name - The name of the connection to update.

Request Lenses

updateConnection_authParameters :: Lens' UpdateConnection (Maybe UpdateConnectionAuthRequestParameters)

The authorization parameters to use for the connection.

updateConnection_authorizationType :: Lens' UpdateConnection (Maybe ConnectionAuthorizationType)

The type of authorization to use for the connection.

updateConnection_description :: Lens' UpdateConnection (Maybe Text)

A description for the connection.

updateConnection_name :: Lens' UpdateConnection Text

The name of the connection to update.

Destructuring the Response

data UpdateConnectionResponse

See: newUpdateConnectionResponse smart constructor.

Constructors

UpdateConnectionResponse'

Fields

creationTime :: Maybe POSIX

A time stamp for the time that the connection was created.

lastModifiedTime :: Maybe POSIX

A time stamp for the time that the connection was last modified.

lastAuthorizedTime :: Maybe POSIX

A time stamp for the time that the connection was last authorized.

connectionArn :: Maybe Text

The ARN of the connection that was updated.

connectionState :: Maybe ConnectionState

The state of the connection that was updated.

httpStatus :: Int

The response's http status code.

newUpdateConnectionResponse

Arguments

:: Int
$sel:httpStatus:UpdateConnectionResponse'

-> UpdateConnectionResponse

Create a value of UpdateConnectionResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:creationTime:UpdateConnectionResponse', updateConnectionResponse_creationTime - A time stamp for the time that the connection was created.

$sel:lastModifiedTime:UpdateConnectionResponse', updateConnectionResponse_lastModifiedTime - A time stamp for the time that the connection was last modified.

$sel:lastAuthorizedTime:UpdateConnectionResponse', updateConnectionResponse_lastAuthorizedTime - A time stamp for the time that the connection was last authorized.

$sel:connectionArn:UpdateConnectionResponse', updateConnectionResponse_connectionArn - The ARN of the connection that was updated.

$sel:connectionState:UpdateConnectionResponse', updateConnectionResponse_connectionState - The state of the connection that was updated.

$sel:httpStatus:UpdateConnectionResponse', updateConnectionResponse_httpStatus - The response's http status code.

Response Lenses

updateConnectionResponse_creationTime :: Lens' UpdateConnectionResponse (Maybe UTCTime)

A time stamp for the time that the connection was created.

updateConnectionResponse_lastModifiedTime :: Lens' UpdateConnectionResponse (Maybe UTCTime)

A time stamp for the time that the connection was last modified.

updateConnectionResponse_lastAuthorizedTime :: Lens' UpdateConnectionResponse (Maybe UTCTime)

A time stamp for the time that the connection was last authorized.

updateConnectionResponse_connectionArn :: Lens' UpdateConnectionResponse (Maybe Text)

The ARN of the connection that was updated.

updateConnectionResponse_connectionState :: Lens' UpdateConnectionResponse (Maybe ConnectionState)

The state of the connection that was updated.

updateConnectionResponse_httpStatus :: Lens' UpdateConnectionResponse Int

The response's http status code.