Amazonka.CloudWatchLogs.DescribeLogStreams

Description

Lists the log streams for the specified log group. You can list all the log streams or filter the results by prefix. You can also control how the results are ordered.

This operation has a limit of five transactions per second, after which transactions are throttled.

This operation returns paginated results.

Creating a Request

data DescribeLogStreams

See: newDescribeLogStreams smart constructor.

Constructors

DescribeLogStreams'

Fields

orderBy :: Maybe OrderBy

If the value is LogStreamName, the results are ordered by log stream name. If the value is LastEventTime, the results are ordered by the event time. The default value is LogStreamName.

If you order the results by event time, you cannot specify the logStreamNamePrefix parameter.

lastEventTimestamp represents the time of the most recent log event in the log stream in CloudWatch Logs. This number is expressed as the number of milliseconds after Jan 1, 1970 00:00:00 UTC. lastEventTimestamp updates on an eventual consistency basis. It typically updates in less than an hour from ingestion, but in rare situations might take longer.

descending :: Maybe Bool

If the value is true, results are returned in descending order. If the value is to false, results are returned in ascending order. The default value is false.

nextToken :: Maybe Text

The token for the next set of items to return. (You received this token from a previous call.)

logStreamNamePrefix :: Maybe Text

The prefix to match.

If orderBy is LastEventTime, you cannot specify this parameter.

limit :: Maybe Natural

The maximum number of items returned. If you don't specify a value, the default is up to 50 items.

logGroupName :: Text

The name of the log group.

newDescribeLogStreams

Arguments

:: Text
$sel:logGroupName:DescribeLogStreams'

-> DescribeLogStreams

Create a value of DescribeLogStreams with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:orderBy:DescribeLogStreams', describeLogStreams_orderBy - If the value is LogStreamName, the results are ordered by log stream name. If the value is LastEventTime, the results are ordered by the event time. The default value is LogStreamName.

If you order the results by event time, you cannot specify the logStreamNamePrefix parameter.

lastEventTimestamp represents the time of the most recent log event in the log stream in CloudWatch Logs. This number is expressed as the number of milliseconds after Jan 1, 1970 00:00:00 UTC. lastEventTimestamp updates on an eventual consistency basis. It typically updates in less than an hour from ingestion, but in rare situations might take longer.

$sel:descending:DescribeLogStreams', describeLogStreams_descending - If the value is true, results are returned in descending order. If the value is to false, results are returned in ascending order. The default value is false.

$sel:nextToken:DescribeLogStreams', describeLogStreams_nextToken - The token for the next set of items to return. (You received this token from a previous call.)

$sel:logStreamNamePrefix:DescribeLogStreams', describeLogStreams_logStreamNamePrefix - The prefix to match.

If orderBy is LastEventTime, you cannot specify this parameter.

$sel:limit:DescribeLogStreams', describeLogStreams_limit - The maximum number of items returned. If you don't specify a value, the default is up to 50 items.

$sel:logGroupName:DescribeLogStreams', describeLogStreams_logGroupName - The name of the log group.

Request Lenses

describeLogStreams_orderBy :: Lens' DescribeLogStreams (Maybe OrderBy)

If the value is LogStreamName, the results are ordered by log stream name. If the value is LastEventTime, the results are ordered by the event time. The default value is LogStreamName.

If you order the results by event time, you cannot specify the logStreamNamePrefix parameter.

lastEventTimestamp represents the time of the most recent log event in the log stream in CloudWatch Logs. This number is expressed as the number of milliseconds after Jan 1, 1970 00:00:00 UTC. lastEventTimestamp updates on an eventual consistency basis. It typically updates in less than an hour from ingestion, but in rare situations might take longer.

describeLogStreams_descending :: Lens' DescribeLogStreams (Maybe Bool)

If the value is true, results are returned in descending order. If the value is to false, results are returned in ascending order. The default value is false.

describeLogStreams_nextToken :: Lens' DescribeLogStreams (Maybe Text)

The token for the next set of items to return. (You received this token from a previous call.)

describeLogStreams_logStreamNamePrefix :: Lens' DescribeLogStreams (Maybe Text)

The prefix to match.

If orderBy is LastEventTime, you cannot specify this parameter.

describeLogStreams_limit :: Lens' DescribeLogStreams (Maybe Natural)

The maximum number of items returned. If you don't specify a value, the default is up to 50 items.

describeLogStreams_logGroupName :: Lens' DescribeLogStreams Text

The name of the log group.

Destructuring the Response

data DescribeLogStreamsResponse

See: newDescribeLogStreamsResponse smart constructor.

Constructors

DescribeLogStreamsResponse'

Fields

nextToken :: Maybe Text

logStreams :: Maybe [LogStream]

The log streams.

httpStatus :: Int

The response's http status code.

newDescribeLogStreamsResponse

Arguments

:: Int
$sel:httpStatus:DescribeLogStreamsResponse'

-> DescribeLogStreamsResponse

Create a value of DescribeLogStreamsResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:nextToken:DescribeLogStreams', describeLogStreamsResponse_nextToken - Undocumented member.

$sel:logStreams:DescribeLogStreamsResponse', describeLogStreamsResponse_logStreams - The log streams.

$sel:httpStatus:DescribeLogStreamsResponse', describeLogStreamsResponse_httpStatus - The response's http status code.

Response Lenses

describeLogStreamsResponse_nextToken :: Lens' DescribeLogStreamsResponse (Maybe Text)

Undocumented member.

describeLogStreamsResponse_logStreams :: Lens' DescribeLogStreamsResponse (Maybe [LogStream])

The log streams.

describeLogStreamsResponse_httpStatus :: Lens' DescribeLogStreamsResponse Int

The response's http status code.