Amazonka.CloudWatchLogs.PutDestination

Description

Creates or updates a destination. This operation is used only to create destinations for cross-account subscriptions.

A destination encapsulates a physical resource (such as an Amazon Kinesis stream) and enables you to subscribe to a real-time stream of log events for a different account, ingested using PutLogEvents.

Through an access policy, a destination controls what is written to it. By default, PutDestination does not set any access policy with the destination, which means a cross-account user cannot call PutSubscriptionFilter against this destination. To enable this, the destination owner must call PutDestinationPolicy after PutDestination.

To perform a PutDestination operation, you must also have the iam:PassRole permission.

Creating a Request

data PutDestination

See: newPutDestination smart constructor.

Constructors

PutDestination'

Fields

destinationName :: Text

A name for the destination.

targetArn :: Text

The ARN of an Amazon Kinesis stream to which to deliver matching log events.

roleArn :: Text

The ARN of an IAM role that grants CloudWatch Logs permissions to call the Amazon Kinesis PutRecord operation on the destination stream.

newPutDestination

Arguments

:: Text
$sel:destinationName:PutDestination'

-> Text
$sel:targetArn:PutDestination'

-> Text
$sel:roleArn:PutDestination'

-> PutDestination

Create a value of PutDestination with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:destinationName:PutDestination', putDestination_destinationName - A name for the destination.

$sel:targetArn:PutDestination', putDestination_targetArn - The ARN of an Amazon Kinesis stream to which to deliver matching log events.

$sel:roleArn:PutDestination', putDestination_roleArn - The ARN of an IAM role that grants CloudWatch Logs permissions to call the Amazon Kinesis PutRecord operation on the destination stream.

Request Lenses

putDestination_destinationName :: Lens' PutDestination Text

A name for the destination.

putDestination_targetArn :: Lens' PutDestination Text

The ARN of an Amazon Kinesis stream to which to deliver matching log events.

putDestination_roleArn :: Lens' PutDestination Text

The ARN of an IAM role that grants CloudWatch Logs permissions to call the Amazon Kinesis PutRecord operation on the destination stream.

Destructuring the Response

data PutDestinationResponse

See: newPutDestinationResponse smart constructor.

Constructors

PutDestinationResponse'

Fields

destination :: Maybe Destination

The destination.

httpStatus :: Int

The response's http status code.

newPutDestinationResponse

Arguments

:: Int
$sel:httpStatus:PutDestinationResponse'

-> PutDestinationResponse

Create a value of PutDestinationResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:destination:PutDestinationResponse', putDestinationResponse_destination - The destination.

$sel:httpStatus:PutDestinationResponse', putDestinationResponse_httpStatus - The response's http status code.

Response Lenses

putDestinationResponse_destination :: Lens' PutDestinationResponse (Maybe Destination)

The destination.

putDestinationResponse_httpStatus :: Lens' PutDestinationResponse Int

The response's http status code.