Amazonka.CloudWatchLogs.PutRetentionPolicy

Description

Sets the retention of the specified log group. A retention policy allows you to configure the number of days for which to retain log events in the specified log group.

Creating a Request

data PutRetentionPolicy

See: newPutRetentionPolicy smart constructor.

Constructors

PutRetentionPolicy'

Fields

logGroupName :: Text

The name of the log group.

retentionInDays :: Int

newPutRetentionPolicy

Arguments

:: Text
$sel:logGroupName:PutRetentionPolicy'

-> Int
$sel:retentionInDays:PutRetentionPolicy'

-> PutRetentionPolicy

Create a value of PutRetentionPolicy with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:logGroupName:PutRetentionPolicy', putRetentionPolicy_logGroupName - The name of the log group.

$sel:retentionInDays:PutRetentionPolicy', putRetentionPolicy_retentionInDays - Undocumented member.

Request Lenses

putRetentionPolicy_logGroupName :: Lens' PutRetentionPolicy Text

The name of the log group.

putRetentionPolicy_retentionInDays :: Lens' PutRetentionPolicy Int

Undocumented member.

Destructuring the Response

data PutRetentionPolicyResponse

See: newPutRetentionPolicyResponse smart constructor.

Constructors

PutRetentionPolicyResponse'

newPutRetentionPolicyResponse :: PutRetentionPolicyResponse

Create a value of PutRetentionPolicyResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.