Amazonka.CloudWatchLogs.Types

Description

Service Configuration

defaultService :: Service

API version 2014-03-28 of the Amazon CloudWatch Logs SDK configuration.

Errors

_InvalidParameterException :: AsError a => Getting (First ServiceError) a ServiceError

A parameter is specified incorrectly.

_InvalidSequenceTokenException :: AsError a => Getting (First ServiceError) a ServiceError

The sequence token is not valid. You can get the correct sequence token in the expectedSequenceToken field in the InvalidSequenceTokenException message.

_UnrecognizedClientException :: AsError a => Getting (First ServiceError) a ServiceError

The most likely cause is an invalid Amazon Web Services access key ID or secret key.

_ResourceAlreadyExistsException :: AsError a => Getting (First ServiceError) a ServiceError

The specified resource already exists.

_OperationAbortedException :: AsError a => Getting (First ServiceError) a ServiceError

Multiple requests to update the same resource were in conflict.

_MalformedQueryException :: AsError a => Getting (First ServiceError) a ServiceError

The query string is not valid. Details about this error are displayed in a QueryCompileError object. For more information, see QueryCompileError.

For more information about valid query syntax, see CloudWatch Logs Insights Query Syntax.

_ServiceUnavailableException :: AsError a => Getting (First ServiceError) a ServiceError

The service cannot complete the request.

_DataAlreadyAcceptedException :: AsError a => Getting (First ServiceError) a ServiceError

The event was already logged.

_InvalidOperationException :: AsError a => Getting (First ServiceError) a ServiceError

The operation is not valid on the specified resource.

_ResourceNotFoundException :: AsError a => Getting (First ServiceError) a ServiceError

The specified resource does not exist.

_LimitExceededException :: AsError a => Getting (First ServiceError) a ServiceError

You have reached the maximum number of resources that can be created.

Distribution

newtype Distribution

The method used to distribute log data to the destination, which can be either random or grouped by log stream.

Constructors

Distribution'

Fields

fromDistribution :: Text

Bundled Patterns

pattern Distribution_ByLogStream :: Distribution

pattern Distribution_Random :: Distribution

ExportTaskStatusCode

newtype ExportTaskStatusCode

Constructors

ExportTaskStatusCode'

Fields

fromExportTaskStatusCode :: Text

Bundled Patterns

pattern ExportTaskStatusCode_CANCELLED :: ExportTaskStatusCode

pattern ExportTaskStatusCode_COMPLETED :: ExportTaskStatusCode

pattern ExportTaskStatusCode_FAILED :: ExportTaskStatusCode

pattern ExportTaskStatusCode_PENDING :: ExportTaskStatusCode

pattern ExportTaskStatusCode_PENDING_CANCEL :: ExportTaskStatusCode

pattern ExportTaskStatusCode_RUNNING :: ExportTaskStatusCode

OrderBy

newtype OrderBy

Constructors

OrderBy'

Fields

fromOrderBy :: Text

Bundled Patterns

pattern OrderBy_LastEventTime :: OrderBy

pattern OrderBy_LogStreamName :: OrderBy

QueryStatus

newtype QueryStatus

Constructors

QueryStatus'

Fields

fromQueryStatus :: Text

Bundled Patterns

pattern QueryStatus_Cancelled :: QueryStatus

pattern QueryStatus_Complete :: QueryStatus

pattern QueryStatus_Failed :: QueryStatus

pattern QueryStatus_Running :: QueryStatus

pattern QueryStatus_Scheduled :: QueryStatus

pattern QueryStatus_Timeout :: QueryStatus

pattern QueryStatus_Unknown :: QueryStatus

StandardUnit

newtype StandardUnit

Constructors

StandardUnit'

Fields

fromStandardUnit :: Text

Bundled Patterns

pattern StandardUnit_Bits :: StandardUnit

pattern StandardUnit_Bits_Second :: StandardUnit

pattern StandardUnit_Bytes :: StandardUnit

pattern StandardUnit_Bytes_Second :: StandardUnit

pattern StandardUnit_Count :: StandardUnit

pattern StandardUnit_Count_Second :: StandardUnit

pattern StandardUnit_Gigabits :: StandardUnit

pattern StandardUnit_Gigabits_Second :: StandardUnit

pattern StandardUnit_Gigabytes :: StandardUnit

pattern StandardUnit_Gigabytes_Second :: StandardUnit

pattern StandardUnit_Kilobits :: StandardUnit

pattern StandardUnit_Kilobits_Second :: StandardUnit

pattern StandardUnit_Kilobytes :: StandardUnit

pattern StandardUnit_Kilobytes_Second :: StandardUnit

pattern StandardUnit_Megabits :: StandardUnit

pattern StandardUnit_Megabits_Second :: StandardUnit

pattern StandardUnit_Megabytes :: StandardUnit

pattern StandardUnit_Megabytes_Second :: StandardUnit

pattern StandardUnit_Microseconds :: StandardUnit

pattern StandardUnit_Milliseconds :: StandardUnit

pattern StandardUnit_None :: StandardUnit

pattern StandardUnit_Percent :: StandardUnit

pattern StandardUnit_Seconds :: StandardUnit

pattern StandardUnit_Terabits :: StandardUnit

pattern StandardUnit_Terabits_Second :: StandardUnit

pattern StandardUnit_Terabytes :: StandardUnit

pattern StandardUnit_Terabytes_Second :: StandardUnit

Destination

data Destination

Represents a cross-account destination that receives subscription log events.

See: newDestination smart constructor.

Constructors

Destination'

Fields

targetArn :: Maybe Text

The Amazon Resource Name (ARN) of the physical target where the log events are delivered (for example, a Kinesis stream).

creationTime :: Maybe Natural

The creation time of the destination, expressed as the number of milliseconds after Jan 1, 1970 00:00:00 UTC.

arn :: Maybe Text

The ARN of this destination.

accessPolicy :: Maybe Text

An IAM policy document that governs which Amazon Web Services accounts can create subscription filters against this destination.

destinationName :: Maybe Text

The name of the destination.

roleArn :: Maybe Text

A role for impersonation, used when delivering log events to the target.

newDestination :: Destination

Create a value of Destination with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:targetArn:Destination', destination_targetArn - The Amazon Resource Name (ARN) of the physical target where the log events are delivered (for example, a Kinesis stream).

$sel:creationTime:Destination', destination_creationTime - The creation time of the destination, expressed as the number of milliseconds after Jan 1, 1970 00:00:00 UTC.

$sel:arn:Destination', destination_arn - The ARN of this destination.

$sel:accessPolicy:Destination', destination_accessPolicy - An IAM policy document that governs which Amazon Web Services accounts can create subscription filters against this destination.

$sel:destinationName:Destination', destination_destinationName - The name of the destination.

$sel:roleArn:Destination', destination_roleArn - A role for impersonation, used when delivering log events to the target.

destination_targetArn :: Lens' Destination (Maybe Text)

The Amazon Resource Name (ARN) of the physical target where the log events are delivered (for example, a Kinesis stream).

destination_creationTime :: Lens' Destination (Maybe Natural)

The creation time of the destination, expressed as the number of milliseconds after Jan 1, 1970 00:00:00 UTC.

destination_arn :: Lens' Destination (Maybe Text)

The ARN of this destination.

destination_accessPolicy :: Lens' Destination (Maybe Text)

An IAM policy document that governs which Amazon Web Services accounts can create subscription filters against this destination.

destination_destinationName :: Lens' Destination (Maybe Text)

The name of the destination.

destination_roleArn :: Lens' Destination (Maybe Text)

A role for impersonation, used when delivering log events to the target.

ExportTask

data ExportTask

Represents an export task.

See: newExportTask smart constructor.

Constructors

ExportTask'

Fields

destinationPrefix :: Maybe Text

The prefix that was used as the start of Amazon S3 key for every object exported.

destination :: Maybe Text

The name of the S3 bucket to which the log data was exported.

status :: Maybe ExportTaskStatus

The status of the export task.

taskName :: Maybe Text

The name of the export task.

taskId :: Maybe Text

The ID of the export task.

to :: Maybe Natural

The end time, expressed as the number of milliseconds after Jan 1, 1970 00:00:00 UTC. Events with a timestamp later than this time are not exported.

from :: Maybe Natural

The start time, expressed as the number of milliseconds after Jan 1, 1970 00:00:00 UTC. Events with a timestamp before this time are not exported.

logGroupName :: Maybe Text

The name of the log group from which logs data was exported.

executionInfo :: Maybe ExportTaskExecutionInfo

Execution information about the export task.

newExportTask :: ExportTask

Create a value of ExportTask with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:destinationPrefix:ExportTask', exportTask_destinationPrefix - The prefix that was used as the start of Amazon S3 key for every object exported.

$sel:destination:ExportTask', exportTask_destination - The name of the S3 bucket to which the log data was exported.

$sel:status:ExportTask', exportTask_status - The status of the export task.

$sel:taskName:ExportTask', exportTask_taskName - The name of the export task.

$sel:taskId:ExportTask', exportTask_taskId - The ID of the export task.

$sel:to:ExportTask', exportTask_to - The end time, expressed as the number of milliseconds after Jan 1, 1970 00:00:00 UTC. Events with a timestamp later than this time are not exported.

$sel:from:ExportTask', exportTask_from - The start time, expressed as the number of milliseconds after Jan 1, 1970 00:00:00 UTC. Events with a timestamp before this time are not exported.

$sel:logGroupName:ExportTask', exportTask_logGroupName - The name of the log group from which logs data was exported.

$sel:executionInfo:ExportTask', exportTask_executionInfo - Execution information about the export task.

exportTask_destinationPrefix :: Lens' ExportTask (Maybe Text)

The prefix that was used as the start of Amazon S3 key for every object exported.

exportTask_destination :: Lens' ExportTask (Maybe Text)

The name of the S3 bucket to which the log data was exported.

exportTask_status :: Lens' ExportTask (Maybe ExportTaskStatus)

The status of the export task.

exportTask_taskName :: Lens' ExportTask (Maybe Text)

The name of the export task.

exportTask_taskId :: Lens' ExportTask (Maybe Text)

The ID of the export task.

exportTask_to :: Lens' ExportTask (Maybe Natural)

The end time, expressed as the number of milliseconds after Jan 1, 1970 00:00:00 UTC. Events with a timestamp later than this time are not exported.

exportTask_from :: Lens' ExportTask (Maybe Natural)

The start time, expressed as the number of milliseconds after Jan 1, 1970 00:00:00 UTC. Events with a timestamp before this time are not exported.

exportTask_logGroupName :: Lens' ExportTask (Maybe Text)

The name of the log group from which logs data was exported.

exportTask_executionInfo :: Lens' ExportTask (Maybe ExportTaskExecutionInfo)

Execution information about the export task.

ExportTaskExecutionInfo

data ExportTaskExecutionInfo

Represents the status of an export task.

See: newExportTaskExecutionInfo smart constructor.

Constructors

ExportTaskExecutionInfo'

Fields

creationTime :: Maybe Natural

The creation time of the export task, expressed as the number of milliseconds after Jan 1, 1970 00:00:00 UTC.

completionTime :: Maybe Natural

The completion time of the export task, expressed as the number of milliseconds after Jan 1, 1970 00:00:00 UTC.

newExportTaskExecutionInfo :: ExportTaskExecutionInfo

Create a value of ExportTaskExecutionInfo with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:creationTime:ExportTaskExecutionInfo', exportTaskExecutionInfo_creationTime - The creation time of the export task, expressed as the number of milliseconds after Jan 1, 1970 00:00:00 UTC.

$sel:completionTime:ExportTaskExecutionInfo', exportTaskExecutionInfo_completionTime - The completion time of the export task, expressed as the number of milliseconds after Jan 1, 1970 00:00:00 UTC.

exportTaskExecutionInfo_creationTime :: Lens' ExportTaskExecutionInfo (Maybe Natural)

The creation time of the export task, expressed as the number of milliseconds after Jan 1, 1970 00:00:00 UTC.

exportTaskExecutionInfo_completionTime :: Lens' ExportTaskExecutionInfo (Maybe Natural)

The completion time of the export task, expressed as the number of milliseconds after Jan 1, 1970 00:00:00 UTC.

ExportTaskStatus

data ExportTaskStatus

Represents the status of an export task.

See: newExportTaskStatus smart constructor.

Constructors

ExportTaskStatus'

Fields

code :: Maybe ExportTaskStatusCode

The status code of the export task.

message :: Maybe Text

The status message related to the status code.

newExportTaskStatus :: ExportTaskStatus

Create a value of ExportTaskStatus with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:code:ExportTaskStatus', exportTaskStatus_code - The status code of the export task.

$sel:message:ExportTaskStatus', exportTaskStatus_message - The status message related to the status code.

exportTaskStatus_code :: Lens' ExportTaskStatus (Maybe ExportTaskStatusCode)

The status code of the export task.

exportTaskStatus_message :: Lens' ExportTaskStatus (Maybe Text)

The status message related to the status code.

FilteredLogEvent

data FilteredLogEvent

Represents a matched event.

See: newFilteredLogEvent smart constructor.

Constructors

FilteredLogEvent'

Fields

ingestionTime :: Maybe Natural

The time the event was ingested, expressed as the number of milliseconds after Jan 1, 1970 00:00:00 UTC.

logStreamName :: Maybe Text

The name of the log stream to which this event belongs.

message :: Maybe Text

The data contained in the log event.

timestamp :: Maybe Natural

The time the event occurred, expressed as the number of milliseconds after Jan 1, 1970 00:00:00 UTC.

eventId :: Maybe Text

The ID of the event.

newFilteredLogEvent :: FilteredLogEvent

Create a value of FilteredLogEvent with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:ingestionTime:FilteredLogEvent', filteredLogEvent_ingestionTime - The time the event was ingested, expressed as the number of milliseconds after Jan 1, 1970 00:00:00 UTC.

$sel:logStreamName:FilteredLogEvent', filteredLogEvent_logStreamName - The name of the log stream to which this event belongs.

$sel:message:FilteredLogEvent', filteredLogEvent_message - The data contained in the log event.

$sel:timestamp:FilteredLogEvent', filteredLogEvent_timestamp - The time the event occurred, expressed as the number of milliseconds after Jan 1, 1970 00:00:00 UTC.

$sel:eventId:FilteredLogEvent', filteredLogEvent_eventId - The ID of the event.

filteredLogEvent_ingestionTime :: Lens' FilteredLogEvent (Maybe Natural)

The time the event was ingested, expressed as the number of milliseconds after Jan 1, 1970 00:00:00 UTC.

filteredLogEvent_logStreamName :: Lens' FilteredLogEvent (Maybe Text)

The name of the log stream to which this event belongs.

filteredLogEvent_message :: Lens' FilteredLogEvent (Maybe Text)

The data contained in the log event.

filteredLogEvent_timestamp :: Lens' FilteredLogEvent (Maybe Natural)

The time the event occurred, expressed as the number of milliseconds after Jan 1, 1970 00:00:00 UTC.

filteredLogEvent_eventId :: Lens' FilteredLogEvent (Maybe Text)

The ID of the event.

InputLogEvent

data InputLogEvent

Represents a log event, which is a record of activity that was recorded by the application or resource being monitored.

See: newInputLogEvent smart constructor.

Constructors

InputLogEvent'

Fields

timestamp :: Natural

The time the event occurred, expressed as the number of milliseconds after Jan 1, 1970 00:00:00 UTC.

message :: Text

The raw event message.

newInputLogEvent

Arguments

:: Natural
$sel:timestamp:InputLogEvent'

-> Text
$sel:message:InputLogEvent'

-> InputLogEvent

Create a value of InputLogEvent with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:timestamp:InputLogEvent', inputLogEvent_timestamp - The time the event occurred, expressed as the number of milliseconds after Jan 1, 1970 00:00:00 UTC.

$sel:message:InputLogEvent', inputLogEvent_message - The raw event message.

inputLogEvent_timestamp :: Lens' InputLogEvent Natural

The time the event occurred, expressed as the number of milliseconds after Jan 1, 1970 00:00:00 UTC.

inputLogEvent_message :: Lens' InputLogEvent Text

The raw event message.

LogGroup

data LogGroup

Represents a log group.

See: newLogGroup smart constructor.

Constructors

LogGroup'

Fields

creationTime :: Maybe Natural

The creation time of the log group, expressed as the number of milliseconds after Jan 1, 1970 00:00:00 UTC.

metricFilterCount :: Maybe Int

The number of metric filters.

arn :: Maybe Text

The Amazon Resource Name (ARN) of the log group.

logGroupName :: Maybe Text

The name of the log group.

retentionInDays :: Maybe Int

kmsKeyId :: Maybe Text

The Amazon Resource Name (ARN) of the CMK to use when encrypting log data.

storedBytes :: Maybe Natural

The number of bytes stored.

newLogGroup :: LogGroup

Create a value of LogGroup with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:creationTime:LogGroup', logGroup_creationTime - The creation time of the log group, expressed as the number of milliseconds after Jan 1, 1970 00:00:00 UTC.

$sel:metricFilterCount:LogGroup', logGroup_metricFilterCount - The number of metric filters.

$sel:arn:LogGroup', logGroup_arn - The Amazon Resource Name (ARN) of the log group.

$sel:logGroupName:LogGroup', logGroup_logGroupName - The name of the log group.

$sel:retentionInDays:LogGroup', logGroup_retentionInDays - Undocumented member.

$sel:kmsKeyId:LogGroup', logGroup_kmsKeyId - The Amazon Resource Name (ARN) of the CMK to use when encrypting log data.

$sel:storedBytes:LogGroup', logGroup_storedBytes - The number of bytes stored.

logGroup_creationTime :: Lens' LogGroup (Maybe Natural)

The creation time of the log group, expressed as the number of milliseconds after Jan 1, 1970 00:00:00 UTC.

logGroup_metricFilterCount :: Lens' LogGroup (Maybe Int)

The number of metric filters.

logGroup_arn :: Lens' LogGroup (Maybe Text)

The Amazon Resource Name (ARN) of the log group.

logGroup_logGroupName :: Lens' LogGroup (Maybe Text)

The name of the log group.

logGroup_retentionInDays :: Lens' LogGroup (Maybe Int)

Undocumented member.

logGroup_kmsKeyId :: Lens' LogGroup (Maybe Text)

The Amazon Resource Name (ARN) of the CMK to use when encrypting log data.

logGroup_storedBytes :: Lens' LogGroup (Maybe Natural)

The number of bytes stored.

LogGroupField

data LogGroupField

The fields contained in log events found by a GetLogGroupFields operation, along with the percentage of queried log events in which each field appears.

See: newLogGroupField smart constructor.

Constructors

LogGroupField'

Fields

percent :: Maybe Natural

The percentage of log events queried that contained the field.

name :: Maybe Text

The name of a log field.

newLogGroupField :: LogGroupField

Create a value of LogGroupField with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:percent:LogGroupField', logGroupField_percent - The percentage of log events queried that contained the field.

$sel:name:LogGroupField', logGroupField_name - The name of a log field.

logGroupField_percent :: Lens' LogGroupField (Maybe Natural)

The percentage of log events queried that contained the field.

logGroupField_name :: Lens' LogGroupField (Maybe Text)

The name of a log field.

LogStream

data LogStream

Represents a log stream, which is a sequence of log events from a single emitter of logs.

See: newLogStream smart constructor.

Constructors

LogStream'

Fields

creationTime :: Maybe Natural

The creation time of the stream, expressed as the number of milliseconds after Jan 1, 1970 00:00:00 UTC.

uploadSequenceToken :: Maybe Text

The sequence token.

arn :: Maybe Text

The Amazon Resource Name (ARN) of the log stream.

firstEventTimestamp :: Maybe Natural

The time of the first event, expressed as the number of milliseconds after Jan 1, 1970 00:00:00 UTC.

logStreamName :: Maybe Text

The name of the log stream.

storedBytes :: Maybe Natural

The number of bytes stored.

Important: On June 17, 2019, this parameter was deprecated for log streams, and is always reported as zero. This change applies only to log streams. The storedBytes parameter for log groups is not affected.

lastIngestionTime :: Maybe Natural

The ingestion time, expressed as the number of milliseconds after Jan 1, 1970 00:00:00 UTC.

lastEventTimestamp :: Maybe Natural

The time of the most recent log event in the log stream in CloudWatch Logs. This number is expressed as the number of milliseconds after Jan 1, 1970 00:00:00 UTC. The lastEventTime value updates on an eventual consistency basis. It typically updates in less than an hour from ingestion, but in rare situations might take longer.

newLogStream :: LogStream

Create a value of LogStream with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:creationTime:LogStream', logStream_creationTime - The creation time of the stream, expressed as the number of milliseconds after Jan 1, 1970 00:00:00 UTC.

$sel:uploadSequenceToken:LogStream', logStream_uploadSequenceToken - The sequence token.

$sel:arn:LogStream', logStream_arn - The Amazon Resource Name (ARN) of the log stream.

$sel:firstEventTimestamp:LogStream', logStream_firstEventTimestamp - The time of the first event, expressed as the number of milliseconds after Jan 1, 1970 00:00:00 UTC.

$sel:logStreamName:LogStream', logStream_logStreamName - The name of the log stream.

$sel:storedBytes:LogStream', logStream_storedBytes - The number of bytes stored.

Important: On June 17, 2019, this parameter was deprecated for log streams, and is always reported as zero. This change applies only to log streams. The storedBytes parameter for log groups is not affected.

$sel:lastIngestionTime:LogStream', logStream_lastIngestionTime - The ingestion time, expressed as the number of milliseconds after Jan 1, 1970 00:00:00 UTC.

$sel:lastEventTimestamp:LogStream', logStream_lastEventTimestamp - The time of the most recent log event in the log stream in CloudWatch Logs. This number is expressed as the number of milliseconds after Jan 1, 1970 00:00:00 UTC. The lastEventTime value updates on an eventual consistency basis. It typically updates in less than an hour from ingestion, but in rare situations might take longer.

logStream_creationTime :: Lens' LogStream (Maybe Natural)

The creation time of the stream, expressed as the number of milliseconds after Jan 1, 1970 00:00:00 UTC.

logStream_uploadSequenceToken :: Lens' LogStream (Maybe Text)

The sequence token.

logStream_arn :: Lens' LogStream (Maybe Text)

The Amazon Resource Name (ARN) of the log stream.

logStream_firstEventTimestamp :: Lens' LogStream (Maybe Natural)

The time of the first event, expressed as the number of milliseconds after Jan 1, 1970 00:00:00 UTC.

logStream_logStreamName :: Lens' LogStream (Maybe Text)

The name of the log stream.

logStream_storedBytes :: Lens' LogStream (Maybe Natural)

The number of bytes stored.

Important: On June 17, 2019, this parameter was deprecated for log streams, and is always reported as zero. This change applies only to log streams. The storedBytes parameter for log groups is not affected.

logStream_lastIngestionTime :: Lens' LogStream (Maybe Natural)

The ingestion time, expressed as the number of milliseconds after Jan 1, 1970 00:00:00 UTC.

logStream_lastEventTimestamp :: Lens' LogStream (Maybe Natural)

The time of the most recent log event in the log stream in CloudWatch Logs. This number is expressed as the number of milliseconds after Jan 1, 1970 00:00:00 UTC. The lastEventTime value updates on an eventual consistency basis. It typically updates in less than an hour from ingestion, but in rare situations might take longer.

MetricFilter

data MetricFilter

Metric filters express how CloudWatch Logs would extract metric observations from ingested log events and transform them into metric data in a CloudWatch metric.

See: newMetricFilter smart constructor.

Constructors

MetricFilter'

Fields

creationTime :: Maybe Natural

The creation time of the metric filter, expressed as the number of milliseconds after Jan 1, 1970 00:00:00 UTC.

filterName :: Maybe Text

The name of the metric filter.

logGroupName :: Maybe Text

The name of the log group.

filterPattern :: Maybe Text

metricTransformations :: Maybe (NonEmpty MetricTransformation)

The metric transformations.

newMetricFilter :: MetricFilter

Create a value of MetricFilter with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:creationTime:MetricFilter', metricFilter_creationTime - The creation time of the metric filter, expressed as the number of milliseconds after Jan 1, 1970 00:00:00 UTC.

$sel:filterName:MetricFilter', metricFilter_filterName - The name of the metric filter.

$sel:logGroupName:MetricFilter', metricFilter_logGroupName - The name of the log group.

$sel:filterPattern:MetricFilter', metricFilter_filterPattern - Undocumented member.

$sel:metricTransformations:MetricFilter', metricFilter_metricTransformations - The metric transformations.

metricFilter_creationTime :: Lens' MetricFilter (Maybe Natural)

The creation time of the metric filter, expressed as the number of milliseconds after Jan 1, 1970 00:00:00 UTC.

metricFilter_filterName :: Lens' MetricFilter (Maybe Text)

The name of the metric filter.

metricFilter_logGroupName :: Lens' MetricFilter (Maybe Text)

The name of the log group.

metricFilter_filterPattern :: Lens' MetricFilter (Maybe Text)

Undocumented member.

metricFilter_metricTransformations :: Lens' MetricFilter (Maybe (NonEmpty MetricTransformation))

The metric transformations.

MetricFilterMatchRecord

data MetricFilterMatchRecord

Represents a matched event.

See: newMetricFilterMatchRecord smart constructor.

Constructors

MetricFilterMatchRecord'

Fields

extractedValues :: Maybe (HashMap Text Text)

The values extracted from the event data by the filter.

eventNumber :: Maybe Integer

The event number.

eventMessage :: Maybe Text

The raw event data.

newMetricFilterMatchRecord :: MetricFilterMatchRecord

Create a value of MetricFilterMatchRecord with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:extractedValues:MetricFilterMatchRecord', metricFilterMatchRecord_extractedValues - The values extracted from the event data by the filter.

$sel:eventNumber:MetricFilterMatchRecord', metricFilterMatchRecord_eventNumber - The event number.

$sel:eventMessage:MetricFilterMatchRecord', metricFilterMatchRecord_eventMessage - The raw event data.

metricFilterMatchRecord_extractedValues :: Lens' MetricFilterMatchRecord (Maybe (HashMap Text Text))

The values extracted from the event data by the filter.

metricFilterMatchRecord_eventNumber :: Lens' MetricFilterMatchRecord (Maybe Integer)

The event number.

metricFilterMatchRecord_eventMessage :: Lens' MetricFilterMatchRecord (Maybe Text)

The raw event data.

MetricTransformation

data MetricTransformation

Indicates how to transform ingested log events to metric data in a CloudWatch metric.

See: newMetricTransformation smart constructor.

Constructors

MetricTransformation'

Fields

defaultValue :: Maybe Double

(Optional) The value to emit when a filter pattern does not match a log event. This value can be null.

dimensions :: Maybe (HashMap Text Text)

The fields to use as dimensions for the metric. One metric filter can include as many as three dimensions.

Metrics extracted from log events are charged as custom metrics. To prevent unexpected high charges, do not specify high-cardinality fields such as IPAddress or requestID as dimensions. Each different value found for a dimension is treated as a separate metric and accrues charges as a separate custom metric.

To help prevent accidental high charges, Amazon disables a metric filter if it generates 1000 different name/value pairs for the dimensions that you have specified within a certain amount of time.

You can also set up a billing alarm to alert you if your charges are higher than expected. For more information, see Creating a Billing Alarm to Monitor Your Estimated Amazon Web Services Charges.

unit :: Maybe StandardUnit

The unit to assign to the metric. If you omit this, the unit is set as None.

metricName :: Text

The name of the CloudWatch metric.

metricNamespace :: Text

A custom namespace to contain your metric in CloudWatch. Use namespaces to group together metrics that are similar. For more information, see Namespaces.

metricValue :: Text

The value to publish to the CloudWatch metric when a filter pattern matches a log event.

newMetricTransformation

Arguments

:: Text
$sel:metricName:MetricTransformation'

-> Text
$sel:metricNamespace:MetricTransformation'

-> Text
$sel:metricValue:MetricTransformation'

-> MetricTransformation

Create a value of MetricTransformation with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:defaultValue:MetricTransformation', metricTransformation_defaultValue - (Optional) The value to emit when a filter pattern does not match a log event. This value can be null.

$sel:dimensions:MetricTransformation', metricTransformation_dimensions - The fields to use as dimensions for the metric. One metric filter can include as many as three dimensions.

Metrics extracted from log events are charged as custom metrics. To prevent unexpected high charges, do not specify high-cardinality fields such as IPAddress or requestID as dimensions. Each different value found for a dimension is treated as a separate metric and accrues charges as a separate custom metric.

To help prevent accidental high charges, Amazon disables a metric filter if it generates 1000 different name/value pairs for the dimensions that you have specified within a certain amount of time.

You can also set up a billing alarm to alert you if your charges are higher than expected. For more information, see Creating a Billing Alarm to Monitor Your Estimated Amazon Web Services Charges.

$sel:unit:MetricTransformation', metricTransformation_unit - The unit to assign to the metric. If you omit this, the unit is set as None.

$sel:metricName:MetricTransformation', metricTransformation_metricName - The name of the CloudWatch metric.

$sel:metricNamespace:MetricTransformation', metricTransformation_metricNamespace - A custom namespace to contain your metric in CloudWatch. Use namespaces to group together metrics that are similar. For more information, see Namespaces.

$sel:metricValue:MetricTransformation', metricTransformation_metricValue - The value to publish to the CloudWatch metric when a filter pattern matches a log event.

metricTransformation_defaultValue :: Lens' MetricTransformation (Maybe Double)

(Optional) The value to emit when a filter pattern does not match a log event. This value can be null.

metricTransformation_dimensions :: Lens' MetricTransformation (Maybe (HashMap Text Text))

The fields to use as dimensions for the metric. One metric filter can include as many as three dimensions.

Metrics extracted from log events are charged as custom metrics. To prevent unexpected high charges, do not specify high-cardinality fields such as IPAddress or requestID as dimensions. Each different value found for a dimension is treated as a separate metric and accrues charges as a separate custom metric.

To help prevent accidental high charges, Amazon disables a metric filter if it generates 1000 different name/value pairs for the dimensions that you have specified within a certain amount of time.

You can also set up a billing alarm to alert you if your charges are higher than expected. For more information, see Creating a Billing Alarm to Monitor Your Estimated Amazon Web Services Charges.

metricTransformation_unit :: Lens' MetricTransformation (Maybe StandardUnit)

The unit to assign to the metric. If you omit this, the unit is set as None.

metricTransformation_metricName :: Lens' MetricTransformation Text

The name of the CloudWatch metric.

metricTransformation_metricNamespace :: Lens' MetricTransformation Text

A custom namespace to contain your metric in CloudWatch. Use namespaces to group together metrics that are similar. For more information, see Namespaces.

metricTransformation_metricValue :: Lens' MetricTransformation Text

The value to publish to the CloudWatch metric when a filter pattern matches a log event.

OutputLogEvent

data OutputLogEvent

Represents a log event.

See: newOutputLogEvent smart constructor.

Constructors

OutputLogEvent'

Fields

ingestionTime :: Maybe Natural

The time the event was ingested, expressed as the number of milliseconds after Jan 1, 1970 00:00:00 UTC.

message :: Maybe Text

The data contained in the log event.

timestamp :: Maybe Natural

The time the event occurred, expressed as the number of milliseconds after Jan 1, 1970 00:00:00 UTC.

newOutputLogEvent :: OutputLogEvent

Create a value of OutputLogEvent with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:ingestionTime:OutputLogEvent', outputLogEvent_ingestionTime - The time the event was ingested, expressed as the number of milliseconds after Jan 1, 1970 00:00:00 UTC.

$sel:message:OutputLogEvent', outputLogEvent_message - The data contained in the log event.

$sel:timestamp:OutputLogEvent', outputLogEvent_timestamp - The time the event occurred, expressed as the number of milliseconds after Jan 1, 1970 00:00:00 UTC.

outputLogEvent_ingestionTime :: Lens' OutputLogEvent (Maybe Natural)

The time the event was ingested, expressed as the number of milliseconds after Jan 1, 1970 00:00:00 UTC.

outputLogEvent_message :: Lens' OutputLogEvent (Maybe Text)

The data contained in the log event.

outputLogEvent_timestamp :: Lens' OutputLogEvent (Maybe Natural)

The time the event occurred, expressed as the number of milliseconds after Jan 1, 1970 00:00:00 UTC.

QueryDefinition

data QueryDefinition

This structure contains details about a saved CloudWatch Logs Insights query definition.

See: newQueryDefinition smart constructor.

Constructors

QueryDefinition'

Fields

logGroupNames :: Maybe [Text]

If this query definition contains a list of log groups that it is limited to, that list appears here.

queryDefinitionId :: Maybe Text

The unique ID of the query definition.

name :: Maybe Text

The name of the query definition.

queryString :: Maybe Text

The query string to use for this definition. For more information, see CloudWatch Logs Insights Query Syntax.

lastModified :: Maybe Natural

The date that the query definition was most recently modified.

newQueryDefinition :: QueryDefinition

Create a value of QueryDefinition with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:logGroupNames:QueryDefinition', queryDefinition_logGroupNames - If this query definition contains a list of log groups that it is limited to, that list appears here.

$sel:queryDefinitionId:QueryDefinition', queryDefinition_queryDefinitionId - The unique ID of the query definition.

$sel:name:QueryDefinition', queryDefinition_name - The name of the query definition.

$sel:queryString:QueryDefinition', queryDefinition_queryString - The query string to use for this definition. For more information, see CloudWatch Logs Insights Query Syntax.

$sel:lastModified:QueryDefinition', queryDefinition_lastModified - The date that the query definition was most recently modified.

queryDefinition_logGroupNames :: Lens' QueryDefinition (Maybe [Text])

If this query definition contains a list of log groups that it is limited to, that list appears here.

queryDefinition_queryDefinitionId :: Lens' QueryDefinition (Maybe Text)

The unique ID of the query definition.

queryDefinition_name :: Lens' QueryDefinition (Maybe Text)

The name of the query definition.

queryDefinition_queryString :: Lens' QueryDefinition (Maybe Text)

The query string to use for this definition. For more information, see CloudWatch Logs Insights Query Syntax.

queryDefinition_lastModified :: Lens' QueryDefinition (Maybe Natural)

The date that the query definition was most recently modified.

QueryInfo

data QueryInfo

Information about one CloudWatch Logs Insights query that matches the request in a DescribeQueries operation.

See: newQueryInfo smart constructor.

Constructors

QueryInfo'

Fields

status :: Maybe QueryStatus

The status of this query. Possible values are Cancelled, Complete, Failed, Running, Scheduled, and Unknown.

queryId :: Maybe Text

The unique ID number of this query.

logGroupName :: Maybe Text

The name of the log group scanned by this query.

queryString :: Maybe Text

The query string used in this query.

createTime :: Maybe Natural

The date and time that this query was created.

newQueryInfo :: QueryInfo

Create a value of QueryInfo with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:status:QueryInfo', queryInfo_status - The status of this query. Possible values are Cancelled, Complete, Failed, Running, Scheduled, and Unknown.

$sel:queryId:QueryInfo', queryInfo_queryId - The unique ID number of this query.

$sel:logGroupName:QueryInfo', queryInfo_logGroupName - The name of the log group scanned by this query.

$sel:queryString:QueryInfo', queryInfo_queryString - The query string used in this query.

$sel:createTime:QueryInfo', queryInfo_createTime - The date and time that this query was created.

queryInfo_status :: Lens' QueryInfo (Maybe QueryStatus)

The status of this query. Possible values are Cancelled, Complete, Failed, Running, Scheduled, and Unknown.

queryInfo_queryId :: Lens' QueryInfo (Maybe Text)

The unique ID number of this query.

queryInfo_logGroupName :: Lens' QueryInfo (Maybe Text)

The name of the log group scanned by this query.

queryInfo_queryString :: Lens' QueryInfo (Maybe Text)

The query string used in this query.

queryInfo_createTime :: Lens' QueryInfo (Maybe Natural)

The date and time that this query was created.

QueryStatistics

data QueryStatistics

Contains the number of log events scanned by the query, the number of log events that matched the query criteria, and the total number of bytes in the log events that were scanned.

See: newQueryStatistics smart constructor.

Constructors

QueryStatistics'

Fields

recordsScanned :: Maybe Double

The total number of log events scanned during the query.

bytesScanned :: Maybe Double

The total number of bytes in the log events scanned during the query.

recordsMatched :: Maybe Double

The number of log events that matched the query string.

newQueryStatistics :: QueryStatistics

Create a value of QueryStatistics with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:recordsScanned:QueryStatistics', queryStatistics_recordsScanned - The total number of log events scanned during the query.

$sel:bytesScanned:QueryStatistics', queryStatistics_bytesScanned - The total number of bytes in the log events scanned during the query.

$sel:recordsMatched:QueryStatistics', queryStatistics_recordsMatched - The number of log events that matched the query string.

queryStatistics_recordsScanned :: Lens' QueryStatistics (Maybe Double)

The total number of log events scanned during the query.

queryStatistics_bytesScanned :: Lens' QueryStatistics (Maybe Double)

The total number of bytes in the log events scanned during the query.

queryStatistics_recordsMatched :: Lens' QueryStatistics (Maybe Double)

The number of log events that matched the query string.

RejectedLogEventsInfo

data RejectedLogEventsInfo

Represents the rejected events.

See: newRejectedLogEventsInfo smart constructor.

Constructors

RejectedLogEventsInfo'

Fields

tooOldLogEventEndIndex :: Maybe Int

The log events that are too old.

tooNewLogEventStartIndex :: Maybe Int

The log events that are too new.

expiredLogEventEndIndex :: Maybe Int

The expired log events.

newRejectedLogEventsInfo :: RejectedLogEventsInfo

Create a value of RejectedLogEventsInfo with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:tooOldLogEventEndIndex:RejectedLogEventsInfo', rejectedLogEventsInfo_tooOldLogEventEndIndex - The log events that are too old.

$sel:tooNewLogEventStartIndex:RejectedLogEventsInfo', rejectedLogEventsInfo_tooNewLogEventStartIndex - The log events that are too new.

$sel:expiredLogEventEndIndex:RejectedLogEventsInfo', rejectedLogEventsInfo_expiredLogEventEndIndex - The expired log events.

rejectedLogEventsInfo_tooOldLogEventEndIndex :: Lens' RejectedLogEventsInfo (Maybe Int)

The log events that are too old.

rejectedLogEventsInfo_tooNewLogEventStartIndex :: Lens' RejectedLogEventsInfo (Maybe Int)

The log events that are too new.

rejectedLogEventsInfo_expiredLogEventEndIndex :: Lens' RejectedLogEventsInfo (Maybe Int)

The expired log events.

ResourcePolicy

data ResourcePolicy

A policy enabling one or more entities to put logs to a log group in this account.

See: newResourcePolicy smart constructor.

Constructors

ResourcePolicy'

Fields

policyName :: Maybe Text

The name of the resource policy.

policyDocument :: Maybe Text

The details of the policy.

lastUpdatedTime :: Maybe Natural

Timestamp showing when this policy was last updated, expressed as the number of milliseconds after Jan 1, 1970 00:00:00 UTC.

newResourcePolicy :: ResourcePolicy

Create a value of ResourcePolicy with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:policyName:ResourcePolicy', resourcePolicy_policyName - The name of the resource policy.

$sel:policyDocument:ResourcePolicy', resourcePolicy_policyDocument - The details of the policy.

$sel:lastUpdatedTime:ResourcePolicy', resourcePolicy_lastUpdatedTime - Timestamp showing when this policy was last updated, expressed as the number of milliseconds after Jan 1, 1970 00:00:00 UTC.

resourcePolicy_policyName :: Lens' ResourcePolicy (Maybe Text)

The name of the resource policy.

resourcePolicy_policyDocument :: Lens' ResourcePolicy (Maybe Text)

The details of the policy.

resourcePolicy_lastUpdatedTime :: Lens' ResourcePolicy (Maybe Natural)

Timestamp showing when this policy was last updated, expressed as the number of milliseconds after Jan 1, 1970 00:00:00 UTC.

ResultField

data ResultField

Contains one field from one log event returned by a CloudWatch Logs Insights query, along with the value of that field.

For more information about the fields that are generated by CloudWatch logs, see Supported Logs and Discovered Fields.

See: newResultField smart constructor.

Constructors

ResultField'

Fields

field :: Maybe Text

The log event field.

value :: Maybe Text

The value of this field.

newResultField :: ResultField

Create a value of ResultField with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:field:ResultField', resultField_field - The log event field.

$sel:value:ResultField', resultField_value - The value of this field.

resultField_field :: Lens' ResultField (Maybe Text)

The log event field.

resultField_value :: Lens' ResultField (Maybe Text)

The value of this field.

SearchedLogStream

data SearchedLogStream

Represents the search status of a log stream.

See: newSearchedLogStream smart constructor.

Constructors

SearchedLogStream'

Fields

logStreamName :: Maybe Text

The name of the log stream.

searchedCompletely :: Maybe Bool

Indicates whether all the events in this log stream were searched.

newSearchedLogStream :: SearchedLogStream

Create a value of SearchedLogStream with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:logStreamName:SearchedLogStream', searchedLogStream_logStreamName - The name of the log stream.

$sel:searchedCompletely:SearchedLogStream', searchedLogStream_searchedCompletely - Indicates whether all the events in this log stream were searched.

searchedLogStream_logStreamName :: Lens' SearchedLogStream (Maybe Text)

The name of the log stream.

searchedLogStream_searchedCompletely :: Lens' SearchedLogStream (Maybe Bool)

Indicates whether all the events in this log stream were searched.

SubscriptionFilter

data SubscriptionFilter

Represents a subscription filter.

See: newSubscriptionFilter smart constructor.

Constructors

SubscriptionFilter'

Fields

creationTime :: Maybe Natural

The creation time of the subscription filter, expressed as the number of milliseconds after Jan 1, 1970 00:00:00 UTC.

filterName :: Maybe Text

The name of the subscription filter.

distribution :: Maybe Distribution

destinationArn :: Maybe Text

The Amazon Resource Name (ARN) of the destination.

logGroupName :: Maybe Text

The name of the log group.

filterPattern :: Maybe Text

roleArn :: Maybe Text

newSubscriptionFilter :: SubscriptionFilter

Create a value of SubscriptionFilter with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:creationTime:SubscriptionFilter', subscriptionFilter_creationTime - The creation time of the subscription filter, expressed as the number of milliseconds after Jan 1, 1970 00:00:00 UTC.

$sel:filterName:SubscriptionFilter', subscriptionFilter_filterName - The name of the subscription filter.

$sel:distribution:SubscriptionFilter', subscriptionFilter_distribution - Undocumented member.

$sel:destinationArn:SubscriptionFilter', subscriptionFilter_destinationArn - The Amazon Resource Name (ARN) of the destination.

$sel:logGroupName:SubscriptionFilter', subscriptionFilter_logGroupName - The name of the log group.

$sel:filterPattern:SubscriptionFilter', subscriptionFilter_filterPattern - Undocumented member.

$sel:roleArn:SubscriptionFilter', subscriptionFilter_roleArn -

subscriptionFilter_creationTime :: Lens' SubscriptionFilter (Maybe Natural)

The creation time of the subscription filter, expressed as the number of milliseconds after Jan 1, 1970 00:00:00 UTC.

subscriptionFilter_filterName :: Lens' SubscriptionFilter (Maybe Text)

The name of the subscription filter.

subscriptionFilter_distribution :: Lens' SubscriptionFilter (Maybe Distribution)

Undocumented member.

subscriptionFilter_destinationArn :: Lens' SubscriptionFilter (Maybe Text)

The Amazon Resource Name (ARN) of the destination.

subscriptionFilter_logGroupName :: Lens' SubscriptionFilter (Maybe Text)

The name of the log group.

subscriptionFilter_filterPattern :: Lens' SubscriptionFilter (Maybe Text)

Undocumented member.

subscriptionFilter_roleArn :: Lens' SubscriptionFilter (Maybe Text)