Amazonka.CloudWatchLogs.Types.ExportTask

Description

Documentation

data ExportTask

Represents an export task.

See: newExportTask smart constructor.

Constructors

ExportTask'

Fields

destinationPrefix :: Maybe Text

The prefix that was used as the start of Amazon S3 key for every object exported.

destination :: Maybe Text

The name of the S3 bucket to which the log data was exported.

status :: Maybe ExportTaskStatus

The status of the export task.

taskName :: Maybe Text

The name of the export task.

taskId :: Maybe Text

The ID of the export task.

to :: Maybe Natural

The end time, expressed as the number of milliseconds after Jan 1, 1970 00:00:00 UTC. Events with a timestamp later than this time are not exported.

from :: Maybe Natural

The start time, expressed as the number of milliseconds after Jan 1, 1970 00:00:00 UTC. Events with a timestamp before this time are not exported.

logGroupName :: Maybe Text

The name of the log group from which logs data was exported.

executionInfo :: Maybe ExportTaskExecutionInfo

Execution information about the export task.

newExportTask :: ExportTask

Create a value of ExportTask with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:destinationPrefix:ExportTask', exportTask_destinationPrefix - The prefix that was used as the start of Amazon S3 key for every object exported.

$sel:destination:ExportTask', exportTask_destination - The name of the S3 bucket to which the log data was exported.

$sel:status:ExportTask', exportTask_status - The status of the export task.

$sel:taskName:ExportTask', exportTask_taskName - The name of the export task.

$sel:taskId:ExportTask', exportTask_taskId - The ID of the export task.

$sel:to:ExportTask', exportTask_to - The end time, expressed as the number of milliseconds after Jan 1, 1970 00:00:00 UTC. Events with a timestamp later than this time are not exported.

$sel:from:ExportTask', exportTask_from - The start time, expressed as the number of milliseconds after Jan 1, 1970 00:00:00 UTC. Events with a timestamp before this time are not exported.

$sel:logGroupName:ExportTask', exportTask_logGroupName - The name of the log group from which logs data was exported.

$sel:executionInfo:ExportTask', exportTask_executionInfo - Execution information about the export task.

exportTask_destinationPrefix :: Lens' ExportTask (Maybe Text)

The prefix that was used as the start of Amazon S3 key for every object exported.

exportTask_destination :: Lens' ExportTask (Maybe Text)

The name of the S3 bucket to which the log data was exported.

exportTask_status :: Lens' ExportTask (Maybe ExportTaskStatus)

The status of the export task.

exportTask_taskName :: Lens' ExportTask (Maybe Text)

The name of the export task.

exportTask_taskId :: Lens' ExportTask (Maybe Text)

The ID of the export task.

exportTask_to :: Lens' ExportTask (Maybe Natural)

The end time, expressed as the number of milliseconds after Jan 1, 1970 00:00:00 UTC. Events with a timestamp later than this time are not exported.

exportTask_from :: Lens' ExportTask (Maybe Natural)

The start time, expressed as the number of milliseconds after Jan 1, 1970 00:00:00 UTC. Events with a timestamp before this time are not exported.

exportTask_logGroupName :: Lens' ExportTask (Maybe Text)

The name of the log group from which logs data was exported.

exportTask_executionInfo :: Lens' ExportTask (Maybe ExportTaskExecutionInfo)

Execution information about the export task.