Amazonka.CodeArtifact.CopyPackageVersions

Description

Copies package versions from one repository to another repository in the same domain.

You must specify versions or versionRevisions. You cannot specify both.

Creating a Request

data CopyPackageVersions

See: newCopyPackageVersions smart constructor.

Constructors

CopyPackageVersions'

Fields

allowOverwrite :: Maybe Bool

Set to true to overwrite a package version that already exists in the destination repository. If set to false and the package version already exists in the destination repository, the package version is returned in the failedVersions field of the response with an ALREADY_EXISTS error code.

versionRevisions :: Maybe (HashMap Text Text)

A list of key-value pairs. The keys are package versions and the values are package version revisions. A CopyPackageVersion operation succeeds if the specified versions in the source repository match the specified package version revision.

You must specify versions or versionRevisions. You cannot specify both.

namespace :: Maybe Text

The namespace of the package. The package component that specifies its namespace depends on its type. For example:

The namespace of a Maven package is its groupId.

The namespace of an npm package is its scope.

A Python package does not contain a corresponding component, so Python packages do not have a namespace.

domainOwner :: Maybe Text

The 12-digit account number of the AWS account that owns the domain. It does not include dashes or spaces.

versions :: Maybe [Text]

The versions of the package to copy.

You must specify versions or versionRevisions. You cannot specify both.

includeFromUpstream :: Maybe Bool

Set to true to copy packages from repositories that are upstream from the source repository to the destination repository. The default setting is false. For more information, see Working with upstream repositories.

domain :: Text

The name of the domain that contains the source and destination repositories.

sourceRepository :: Text

The name of the repository that contains the package versions to copy.

destinationRepository :: Text

The name of the repository into which package versions are copied.

format :: PackageFormat

The format of the package that is copied. The valid package types are:

npm: A Node Package Manager (npm) package.

pypi: A Python Package Index (PyPI) package.

maven: A Maven package that contains compiled code in a distributable format, such as a JAR file.

package :: Text

The name of the package that is copied.

newCopyPackageVersions

Arguments

:: Text
$sel:domain:CopyPackageVersions'

-> Text
$sel:sourceRepository:CopyPackageVersions'

-> Text
$sel:destinationRepository:CopyPackageVersions'

-> PackageFormat
$sel:format:CopyPackageVersions'

-> Text
$sel:package:CopyPackageVersions'

-> CopyPackageVersions

Create a value of CopyPackageVersions with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:allowOverwrite:CopyPackageVersions', copyPackageVersions_allowOverwrite - Set to true to overwrite a package version that already exists in the destination repository. If set to false and the package version already exists in the destination repository, the package version is returned in the failedVersions field of the response with an ALREADY_EXISTS error code.

$sel:versionRevisions:CopyPackageVersions', copyPackageVersions_versionRevisions - A list of key-value pairs. The keys are package versions and the values are package version revisions. A CopyPackageVersion operation succeeds if the specified versions in the source repository match the specified package version revision.

You must specify versions or versionRevisions. You cannot specify both.

$sel:namespace:CopyPackageVersions', copyPackageVersions_namespace - The namespace of the package. The package component that specifies its namespace depends on its type. For example:

The namespace of a Maven package is its groupId.

The namespace of an npm package is its scope.

A Python package does not contain a corresponding component, so Python packages do not have a namespace.

$sel:domainOwner:CopyPackageVersions', copyPackageVersions_domainOwner - The 12-digit account number of the AWS account that owns the domain. It does not include dashes or spaces.

$sel:versions:CopyPackageVersions', copyPackageVersions_versions - The versions of the package to copy.

You must specify versions or versionRevisions. You cannot specify both.

$sel:includeFromUpstream:CopyPackageVersions', copyPackageVersions_includeFromUpstream - Set to true to copy packages from repositories that are upstream from the source repository to the destination repository. The default setting is false. For more information, see Working with upstream repositories.

$sel:domain:CopyPackageVersions', copyPackageVersions_domain - The name of the domain that contains the source and destination repositories.

$sel:sourceRepository:CopyPackageVersions', copyPackageVersions_sourceRepository - The name of the repository that contains the package versions to copy.

$sel:destinationRepository:CopyPackageVersions', copyPackageVersions_destinationRepository - The name of the repository into which package versions are copied.

$sel:format:CopyPackageVersions', copyPackageVersions_format - The format of the package that is copied. The valid package types are:

npm: A Node Package Manager (npm) package.

pypi: A Python Package Index (PyPI) package.

maven: A Maven package that contains compiled code in a distributable format, such as a JAR file.

$sel:package:CopyPackageVersions', copyPackageVersions_package - The name of the package that is copied.

Request Lenses

copyPackageVersions_allowOverwrite :: Lens' CopyPackageVersions (Maybe Bool)

Set to true to overwrite a package version that already exists in the destination repository. If set to false and the package version already exists in the destination repository, the package version is returned in the failedVersions field of the response with an ALREADY_EXISTS error code.

copyPackageVersions_versionRevisions :: Lens' CopyPackageVersions (Maybe (HashMap Text Text))

A list of key-value pairs. The keys are package versions and the values are package version revisions. A CopyPackageVersion operation succeeds if the specified versions in the source repository match the specified package version revision.

You must specify versions or versionRevisions. You cannot specify both.

copyPackageVersions_namespace :: Lens' CopyPackageVersions (Maybe Text)

The namespace of the package. The package component that specifies its namespace depends on its type. For example:

The namespace of a Maven package is its groupId.

The namespace of an npm package is its scope.

A Python package does not contain a corresponding component, so Python packages do not have a namespace.

copyPackageVersions_domainOwner :: Lens' CopyPackageVersions (Maybe Text)

The 12-digit account number of the AWS account that owns the domain. It does not include dashes or spaces.

copyPackageVersions_versions :: Lens' CopyPackageVersions (Maybe [Text])

The versions of the package to copy.

You must specify versions or versionRevisions. You cannot specify both.

copyPackageVersions_includeFromUpstream :: Lens' CopyPackageVersions (Maybe Bool)

Set to true to copy packages from repositories that are upstream from the source repository to the destination repository. The default setting is false. For more information, see Working with upstream repositories.

copyPackageVersions_domain :: Lens' CopyPackageVersions Text

The name of the domain that contains the source and destination repositories.

copyPackageVersions_sourceRepository :: Lens' CopyPackageVersions Text

The name of the repository that contains the package versions to copy.

copyPackageVersions_destinationRepository :: Lens' CopyPackageVersions Text

The name of the repository into which package versions are copied.

copyPackageVersions_format :: Lens' CopyPackageVersions PackageFormat

The format of the package that is copied. The valid package types are:

npm: A Node Package Manager (npm) package.

pypi: A Python Package Index (PyPI) package.

maven: A Maven package that contains compiled code in a distributable format, such as a JAR file.

copyPackageVersions_package :: Lens' CopyPackageVersions Text

The name of the package that is copied.

Destructuring the Response

data CopyPackageVersionsResponse

See: newCopyPackageVersionsResponse smart constructor.

Constructors

CopyPackageVersionsResponse'

Fields

failedVersions :: Maybe (HashMap Text PackageVersionError)

A map of package versions that failed to copy and their error codes. The possible error codes are in the PackageVersionError data type. They are:

ALREADY_EXISTS

MISMATCHED_REVISION

MISMATCHED_STATUS

NOT_ALLOWED

NOT_FOUND

SKIPPED

successfulVersions :: Maybe (HashMap Text SuccessfulPackageVersionInfo)

A list of the package versions that were successfully copied to your repository.

httpStatus :: Int

The response's http status code.

newCopyPackageVersionsResponse

Arguments

:: Int
$sel:httpStatus:CopyPackageVersionsResponse'

-> CopyPackageVersionsResponse

Create a value of CopyPackageVersionsResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:failedVersions:CopyPackageVersionsResponse', copyPackageVersionsResponse_failedVersions - A map of package versions that failed to copy and their error codes. The possible error codes are in the PackageVersionError data type. They are:

ALREADY_EXISTS

MISMATCHED_REVISION

MISMATCHED_STATUS

NOT_ALLOWED

NOT_FOUND

SKIPPED

$sel:successfulVersions:CopyPackageVersionsResponse', copyPackageVersionsResponse_successfulVersions - A list of the package versions that were successfully copied to your repository.

$sel:httpStatus:CopyPackageVersionsResponse', copyPackageVersionsResponse_httpStatus - The response's http status code.

Response Lenses

copyPackageVersionsResponse_failedVersions :: Lens' CopyPackageVersionsResponse (Maybe (HashMap Text PackageVersionError))

A map of package versions that failed to copy and their error codes. The possible error codes are in the PackageVersionError data type. They are:

ALREADY_EXISTS

MISMATCHED_REVISION

MISMATCHED_STATUS

NOT_ALLOWED

NOT_FOUND

SKIPPED

copyPackageVersionsResponse_successfulVersions :: Lens' CopyPackageVersionsResponse (Maybe (HashMap Text SuccessfulPackageVersionInfo))

A list of the package versions that were successfully copied to your repository.

copyPackageVersionsResponse_httpStatus :: Lens' CopyPackageVersionsResponse Int

The response's http status code.