Amazonka.CodeArtifact.CreateDomain

Description

Creates a domain. CodeArtifact domains make it easier to manage multiple repositories across an organization. You can use a domain to apply permissions across many repositories owned by different AWS accounts. An asset is stored only once in a domain, even if it's in multiple repositories.

Although you can have multiple domains, we recommend a single production domain that contains all published artifacts so that your development teams can find and share packages. You can use a second pre-production domain to test changes to the production domain configuration.

Creating a Request

data CreateDomain

See: newCreateDomain smart constructor.

Constructors

CreateDomain'

Fields

encryptionKey :: Maybe Text

The encryption key for the domain. This is used to encrypt content stored in a domain. An encryption key can be a key ID, a key Amazon Resource Name (ARN), a key alias, or a key alias ARN. To specify an encryptionKey, your IAM role must have kms:DescribeKey and kms:CreateGrant permissions on the encryption key that is used. For more information, see DescribeKey in the AWS Key Management Service API Reference and AWS KMS API Permissions Reference in the AWS Key Management Service Developer Guide.

CodeArtifact supports only symmetric CMKs. Do not associate an asymmetric CMK with your domain. For more information, see Using symmetric and asymmetric keys in the AWS Key Management Service Developer Guide.

tags :: Maybe [Tag]

One or more tag key-value pairs for the domain.

domain :: Text

The name of the domain to create. All domain names in an AWS Region that are in the same AWS account must be unique. The domain name is used as the prefix in DNS hostnames. Do not use sensitive information in a domain name because it is publicly discoverable.

newCreateDomain

Arguments

:: Text
$sel:domain:CreateDomain'

-> CreateDomain

Create a value of CreateDomain with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:encryptionKey:CreateDomain', createDomain_encryptionKey - The encryption key for the domain. This is used to encrypt content stored in a domain. An encryption key can be a key ID, a key Amazon Resource Name (ARN), a key alias, or a key alias ARN. To specify an encryptionKey, your IAM role must have kms:DescribeKey and kms:CreateGrant permissions on the encryption key that is used. For more information, see DescribeKey in the AWS Key Management Service API Reference and AWS KMS API Permissions Reference in the AWS Key Management Service Developer Guide.

CodeArtifact supports only symmetric CMKs. Do not associate an asymmetric CMK with your domain. For more information, see Using symmetric and asymmetric keys in the AWS Key Management Service Developer Guide.

$sel:tags:CreateDomain', createDomain_tags - One or more tag key-value pairs for the domain.

$sel:domain:CreateDomain', createDomain_domain - The name of the domain to create. All domain names in an AWS Region that are in the same AWS account must be unique. The domain name is used as the prefix in DNS hostnames. Do not use sensitive information in a domain name because it is publicly discoverable.

Request Lenses

createDomain_encryptionKey :: Lens' CreateDomain (Maybe Text)

The encryption key for the domain. This is used to encrypt content stored in a domain. An encryption key can be a key ID, a key Amazon Resource Name (ARN), a key alias, or a key alias ARN. To specify an encryptionKey, your IAM role must have kms:DescribeKey and kms:CreateGrant permissions on the encryption key that is used. For more information, see DescribeKey in the AWS Key Management Service API Reference and AWS KMS API Permissions Reference in the AWS Key Management Service Developer Guide.

CodeArtifact supports only symmetric CMKs. Do not associate an asymmetric CMK with your domain. For more information, see Using symmetric and asymmetric keys in the AWS Key Management Service Developer Guide.

createDomain_tags :: Lens' CreateDomain (Maybe [Tag])

One or more tag key-value pairs for the domain.

createDomain_domain :: Lens' CreateDomain Text

The name of the domain to create. All domain names in an AWS Region that are in the same AWS account must be unique. The domain name is used as the prefix in DNS hostnames. Do not use sensitive information in a domain name because it is publicly discoverable.

Destructuring the Response

data CreateDomainResponse

See: newCreateDomainResponse smart constructor.

Constructors

CreateDomainResponse'

Fields

domain :: Maybe DomainDescription

Contains information about the created domain after processing the request.

httpStatus :: Int

The response's http status code.

newCreateDomainResponse

Arguments

:: Int
$sel:httpStatus:CreateDomainResponse'

-> CreateDomainResponse

Create a value of CreateDomainResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:domain:CreateDomain', createDomainResponse_domain - Contains information about the created domain after processing the request.

$sel:httpStatus:CreateDomainResponse', createDomainResponse_httpStatus - The response's http status code.

Response Lenses

createDomainResponse_domain :: Lens' CreateDomainResponse (Maybe DomainDescription)

Contains information about the created domain after processing the request.

createDomainResponse_httpStatus :: Lens' CreateDomainResponse Int

The response's http status code.