Amazonka.CodeArtifact.DescribePackageVersion

Description

Returns a PackageVersionDescription object that contains information about the requested package version.

Creating a Request

data DescribePackageVersion

See: newDescribePackageVersion smart constructor.

Constructors

DescribePackageVersion'

Fields

namespace :: Maybe Text

The namespace of the package. The package component that specifies its namespace depends on its type. For example:

The namespace of a Maven package is its groupId.

The namespace of an npm package is its scope.

A Python package does not contain a corresponding component, so Python packages do not have a namespace.

domainOwner :: Maybe Text

The 12-digit account number of the AWS account that owns the domain. It does not include dashes or spaces.

domain :: Text

The name of the domain that contains the repository that contains the package version.

repository :: Text

The name of the repository that contains the package version.

format :: PackageFormat

A format that specifies the type of the requested package version. The valid values are:

npm

pypi

maven

package :: Text

The name of the requested package version.

packageVersion :: Text

A string that contains the package version (for example, 3.5.2).

newDescribePackageVersion

Arguments

:: Text
$sel:domain:DescribePackageVersion'

-> Text
$sel:repository:DescribePackageVersion'

-> PackageFormat
$sel:format:DescribePackageVersion'

-> Text
$sel:package:DescribePackageVersion'

-> Text
$sel:packageVersion:DescribePackageVersion'

-> DescribePackageVersion

Create a value of DescribePackageVersion with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:namespace:DescribePackageVersion', describePackageVersion_namespace - The namespace of the package. The package component that specifies its namespace depends on its type. For example:

The namespace of a Maven package is its groupId.

The namespace of an npm package is its scope.

A Python package does not contain a corresponding component, so Python packages do not have a namespace.

$sel:domainOwner:DescribePackageVersion', describePackageVersion_domainOwner - The 12-digit account number of the AWS account that owns the domain. It does not include dashes or spaces.

$sel:domain:DescribePackageVersion', describePackageVersion_domain - The name of the domain that contains the repository that contains the package version.

$sel:repository:DescribePackageVersion', describePackageVersion_repository - The name of the repository that contains the package version.

$sel:format:DescribePackageVersion', describePackageVersion_format - A format that specifies the type of the requested package version. The valid values are:

npm

pypi

maven

$sel:package:DescribePackageVersion', describePackageVersion_package - The name of the requested package version.

$sel:packageVersion:DescribePackageVersion', describePackageVersion_packageVersion - A string that contains the package version (for example, 3.5.2).

Request Lenses

describePackageVersion_namespace :: Lens' DescribePackageVersion (Maybe Text)

The namespace of the package. The package component that specifies its namespace depends on its type. For example:

The namespace of a Maven package is its groupId.

The namespace of an npm package is its scope.

A Python package does not contain a corresponding component, so Python packages do not have a namespace.

describePackageVersion_domainOwner :: Lens' DescribePackageVersion (Maybe Text)

The 12-digit account number of the AWS account that owns the domain. It does not include dashes or spaces.

describePackageVersion_domain :: Lens' DescribePackageVersion Text

The name of the domain that contains the repository that contains the package version.

describePackageVersion_repository :: Lens' DescribePackageVersion Text

The name of the repository that contains the package version.

describePackageVersion_format :: Lens' DescribePackageVersion PackageFormat

A format that specifies the type of the requested package version. The valid values are:

npm

pypi

maven

describePackageVersion_package :: Lens' DescribePackageVersion Text

The name of the requested package version.

describePackageVersion_packageVersion :: Lens' DescribePackageVersion Text

A string that contains the package version (for example, 3.5.2).

Destructuring the Response

data DescribePackageVersionResponse

See: newDescribePackageVersionResponse smart constructor.

Constructors

DescribePackageVersionResponse'

Fields

httpStatus :: Int

The response's http status code.

packageVersion :: PackageVersionDescription

A PackageVersionDescription object that contains information about the requested package version.

newDescribePackageVersionResponse

Arguments

:: Int
$sel:httpStatus:DescribePackageVersionResponse'

-> PackageVersionDescription
$sel:packageVersion:DescribePackageVersion'

-> DescribePackageVersionResponse

Create a value of DescribePackageVersionResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:httpStatus:DescribePackageVersionResponse', describePackageVersionResponse_httpStatus - The response's http status code.

$sel:packageVersion:DescribePackageVersion', describePackageVersionResponse_packageVersion - A PackageVersionDescription object that contains information about the requested package version.

Response Lenses

describePackageVersionResponse_httpStatus :: Lens' DescribePackageVersionResponse Int

The response's http status code.

describePackageVersionResponse_packageVersion :: Lens' DescribePackageVersionResponse PackageVersionDescription

A PackageVersionDescription object that contains information about the requested package version.