Amazonka.CodeArtifact.DisposePackageVersions

Description

Deletes the assets in package versions and sets the package versions' status to Disposed. A disposed package version cannot be restored in your repository because its assets are deleted.

To view all disposed package versions in a repository, use ListPackageVersions and set the status parameter to Disposed.

To view information about a disposed package version, use DescribePackageVersion.

Creating a Request

data DisposePackageVersions

See: newDisposePackageVersions smart constructor.

Constructors

DisposePackageVersions'

Fields

expectedStatus :: Maybe PackageVersionStatus

The expected status of the package version to dispose. Valid values are:

Published

Unfinished

Unlisted

Archived

Disposed

versionRevisions :: Maybe (HashMap Text Text)

The revisions of the package versions you want to dispose.

namespace :: Maybe Text

The namespace of the package. The package component that specifies its namespace depends on its type. For example:

The namespace of a Maven package is its groupId.

The namespace of an npm package is its scope.

A Python package does not contain a corresponding component, so Python packages do not have a namespace.

domainOwner :: Maybe Text

The 12-digit account number of the AWS account that owns the domain. It does not include dashes or spaces.

domain :: Text

The name of the domain that contains the repository you want to dispose.

repository :: Text

The name of the repository that contains the package versions you want to dispose.

format :: PackageFormat

A format that specifies the type of package versions you want to dispose. The valid values are:

npm

pypi

maven

package :: Text

The name of the package with the versions you want to dispose.

versions :: [Text]

The versions of the package you want to dispose.

newDisposePackageVersions

Arguments

:: Text
$sel:domain:DisposePackageVersions'

-> Text
$sel:repository:DisposePackageVersions'

-> PackageFormat
$sel:format:DisposePackageVersions'

-> Text
$sel:package:DisposePackageVersions'

-> DisposePackageVersions

Create a value of DisposePackageVersions with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:expectedStatus:DisposePackageVersions', disposePackageVersions_expectedStatus - The expected status of the package version to dispose. Valid values are:

Published

Unfinished

Unlisted

Archived

Disposed

$sel:versionRevisions:DisposePackageVersions', disposePackageVersions_versionRevisions - The revisions of the package versions you want to dispose.

$sel:namespace:DisposePackageVersions', disposePackageVersions_namespace - The namespace of the package. The package component that specifies its namespace depends on its type. For example:

The namespace of a Maven package is its groupId.

The namespace of an npm package is its scope.

A Python package does not contain a corresponding component, so Python packages do not have a namespace.

$sel:domainOwner:DisposePackageVersions', disposePackageVersions_domainOwner - The 12-digit account number of the AWS account that owns the domain. It does not include dashes or spaces.

$sel:domain:DisposePackageVersions', disposePackageVersions_domain - The name of the domain that contains the repository you want to dispose.

$sel:repository:DisposePackageVersions', disposePackageVersions_repository - The name of the repository that contains the package versions you want to dispose.

$sel:format:DisposePackageVersions', disposePackageVersions_format - A format that specifies the type of package versions you want to dispose. The valid values are:

npm

pypi

maven

$sel:package:DisposePackageVersions', disposePackageVersions_package - The name of the package with the versions you want to dispose.

$sel:versions:DisposePackageVersions', disposePackageVersions_versions - The versions of the package you want to dispose.

Request Lenses

disposePackageVersions_expectedStatus :: Lens' DisposePackageVersions (Maybe PackageVersionStatus)

The expected status of the package version to dispose. Valid values are:

Published

Unfinished

Unlisted

Archived

Disposed

disposePackageVersions_versionRevisions :: Lens' DisposePackageVersions (Maybe (HashMap Text Text))

The revisions of the package versions you want to dispose.

disposePackageVersions_namespace :: Lens' DisposePackageVersions (Maybe Text)

The namespace of the package. The package component that specifies its namespace depends on its type. For example:

The namespace of a Maven package is its groupId.

The namespace of an npm package is its scope.

A Python package does not contain a corresponding component, so Python packages do not have a namespace.

disposePackageVersions_domainOwner :: Lens' DisposePackageVersions (Maybe Text)

The 12-digit account number of the AWS account that owns the domain. It does not include dashes or spaces.

disposePackageVersions_domain :: Lens' DisposePackageVersions Text

The name of the domain that contains the repository you want to dispose.

disposePackageVersions_repository :: Lens' DisposePackageVersions Text

The name of the repository that contains the package versions you want to dispose.

disposePackageVersions_format :: Lens' DisposePackageVersions PackageFormat

A format that specifies the type of package versions you want to dispose. The valid values are:

npm

pypi

maven

disposePackageVersions_package :: Lens' DisposePackageVersions Text

The name of the package with the versions you want to dispose.

disposePackageVersions_versions :: Lens' DisposePackageVersions [Text]

The versions of the package you want to dispose.

Destructuring the Response

data DisposePackageVersionsResponse

See: newDisposePackageVersionsResponse smart constructor.

Constructors

DisposePackageVersionsResponse'

Fields

failedVersions :: Maybe (HashMap Text PackageVersionError)

A PackageVersionError object that contains a map of errors codes for the disposed package versions that failed. The possible error codes are:

ALREADY_EXISTS

MISMATCHED_REVISION

MISMATCHED_STATUS

NOT_ALLOWED

NOT_FOUND

SKIPPED

successfulVersions :: Maybe (HashMap Text SuccessfulPackageVersionInfo)

A list of the package versions that were successfully disposed.

httpStatus :: Int

The response's http status code.

newDisposePackageVersionsResponse

Arguments

:: Int
$sel:httpStatus:DisposePackageVersionsResponse'

-> DisposePackageVersionsResponse

Create a value of DisposePackageVersionsResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:failedVersions:DisposePackageVersionsResponse', disposePackageVersionsResponse_failedVersions - A PackageVersionError object that contains a map of errors codes for the disposed package versions that failed. The possible error codes are:

ALREADY_EXISTS

MISMATCHED_REVISION

MISMATCHED_STATUS

NOT_ALLOWED

NOT_FOUND

SKIPPED

$sel:successfulVersions:DisposePackageVersionsResponse', disposePackageVersionsResponse_successfulVersions - A list of the package versions that were successfully disposed.

$sel:httpStatus:DisposePackageVersionsResponse', disposePackageVersionsResponse_httpStatus - The response's http status code.

Response Lenses

disposePackageVersionsResponse_failedVersions :: Lens' DisposePackageVersionsResponse (Maybe (HashMap Text PackageVersionError))

A PackageVersionError object that contains a map of errors codes for the disposed package versions that failed. The possible error codes are:

ALREADY_EXISTS

MISMATCHED_REVISION

MISMATCHED_STATUS

NOT_ALLOWED

NOT_FOUND

SKIPPED

disposePackageVersionsResponse_successfulVersions :: Lens' DisposePackageVersionsResponse (Maybe (HashMap Text SuccessfulPackageVersionInfo))

A list of the package versions that were successfully disposed.

disposePackageVersionsResponse_httpStatus :: Lens' DisposePackageVersionsResponse Int

The response's http status code.