Amazonka.CodeArtifact.GetRepositoryPermissionsPolicy

Description

Returns the resource policy that is set on a repository.

Creating a Request

data GetRepositoryPermissionsPolicy

See: newGetRepositoryPermissionsPolicy smart constructor.

Constructors

GetRepositoryPermissionsPolicy'

Fields

domainOwner :: Maybe Text

The 12-digit account number of the AWS account that owns the domain. It does not include dashes or spaces.

domain :: Text

The name of the domain containing the repository whose associated resource policy is to be retrieved.

repository :: Text

The name of the repository whose associated resource policy is to be retrieved.

newGetRepositoryPermissionsPolicy

Arguments

:: Text
$sel:domain:GetRepositoryPermissionsPolicy'

-> Text
$sel:repository:GetRepositoryPermissionsPolicy'

-> GetRepositoryPermissionsPolicy

Create a value of GetRepositoryPermissionsPolicy with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:domainOwner:GetRepositoryPermissionsPolicy', getRepositoryPermissionsPolicy_domainOwner - The 12-digit account number of the AWS account that owns the domain. It does not include dashes or spaces.

$sel:domain:GetRepositoryPermissionsPolicy', getRepositoryPermissionsPolicy_domain - The name of the domain containing the repository whose associated resource policy is to be retrieved.

$sel:repository:GetRepositoryPermissionsPolicy', getRepositoryPermissionsPolicy_repository - The name of the repository whose associated resource policy is to be retrieved.

Request Lenses

getRepositoryPermissionsPolicy_domainOwner :: Lens' GetRepositoryPermissionsPolicy (Maybe Text)

The 12-digit account number of the AWS account that owns the domain. It does not include dashes or spaces.

getRepositoryPermissionsPolicy_domain :: Lens' GetRepositoryPermissionsPolicy Text

The name of the domain containing the repository whose associated resource policy is to be retrieved.

getRepositoryPermissionsPolicy_repository :: Lens' GetRepositoryPermissionsPolicy Text

The name of the repository whose associated resource policy is to be retrieved.

Destructuring the Response

data GetRepositoryPermissionsPolicyResponse

See: newGetRepositoryPermissionsPolicyResponse smart constructor.

Constructors

GetRepositoryPermissionsPolicyResponse'

Fields

policy :: Maybe ResourcePolicy

The returned resource policy.

httpStatus :: Int

The response's http status code.

newGetRepositoryPermissionsPolicyResponse

Arguments

:: Int
$sel:httpStatus:GetRepositoryPermissionsPolicyResponse'

-> GetRepositoryPermissionsPolicyResponse

Create a value of GetRepositoryPermissionsPolicyResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:policy:GetRepositoryPermissionsPolicyResponse', getRepositoryPermissionsPolicyResponse_policy - The returned resource policy.

$sel:httpStatus:GetRepositoryPermissionsPolicyResponse', getRepositoryPermissionsPolicyResponse_httpStatus - The response's http status code.

Response Lenses

getRepositoryPermissionsPolicyResponse_policy :: Lens' GetRepositoryPermissionsPolicyResponse (Maybe ResourcePolicy)

The returned resource policy.

getRepositoryPermissionsPolicyResponse_httpStatus :: Lens' GetRepositoryPermissionsPolicyResponse Int

The response's http status code.