Amazonka.CodeArtifact.ListPackageVersionAssets

Description

Returns a list of AssetSummary objects for assets in a package version.

This operation returns paginated results.

Creating a Request

data ListPackageVersionAssets

See: newListPackageVersionAssets smart constructor.

Constructors

ListPackageVersionAssets'

Fields

namespace :: Maybe Text

The namespace of the package. The package component that specifies its namespace depends on its type. For example:

The namespace of a Maven package is its groupId.

The namespace of an npm package is its scope.

A Python package does not contain a corresponding component, so Python packages do not have a namespace.

domainOwner :: Maybe Text

The 12-digit account number of the AWS account that owns the domain. It does not include dashes or spaces.

nextToken :: Maybe Text

The token for the next set of results. Use the value returned in the previous response in the next request to retrieve the next set of results.

maxResults :: Maybe Natural

The maximum number of results to return per page.

domain :: Text

The name of the domain that contains the repository associated with the package version assets.

repository :: Text

The name of the repository that contains the package that contains the returned package version assets.

format :: PackageFormat

The format of the package that contains the returned package version assets. The valid package types are:

npm: A Node Package Manager (npm) package.

pypi: A Python Package Index (PyPI) package.

maven: A Maven package that contains compiled code in a distributable format, such as a JAR file.

package :: Text

The name of the package that contains the returned package version assets.

packageVersion :: Text

A string that contains the package version (for example, 3.5.2).

newListPackageVersionAssets

Arguments

:: Text
$sel:domain:ListPackageVersionAssets'

-> Text
$sel:repository:ListPackageVersionAssets'

-> PackageFormat
$sel:format:ListPackageVersionAssets'

-> Text
$sel:package:ListPackageVersionAssets'

-> Text
$sel:packageVersion:ListPackageVersionAssets'

-> ListPackageVersionAssets

Create a value of ListPackageVersionAssets with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:namespace:ListPackageVersionAssets', listPackageVersionAssets_namespace - The namespace of the package. The package component that specifies its namespace depends on its type. For example:

The namespace of a Maven package is its groupId.

The namespace of an npm package is its scope.

A Python package does not contain a corresponding component, so Python packages do not have a namespace.

$sel:domainOwner:ListPackageVersionAssets', listPackageVersionAssets_domainOwner - The 12-digit account number of the AWS account that owns the domain. It does not include dashes or spaces.

$sel:nextToken:ListPackageVersionAssets', listPackageVersionAssets_nextToken - The token for the next set of results. Use the value returned in the previous response in the next request to retrieve the next set of results.

$sel:maxResults:ListPackageVersionAssets', listPackageVersionAssets_maxResults - The maximum number of results to return per page.

$sel:domain:ListPackageVersionAssets', listPackageVersionAssets_domain - The name of the domain that contains the repository associated with the package version assets.

$sel:repository:ListPackageVersionAssets', listPackageVersionAssets_repository - The name of the repository that contains the package that contains the returned package version assets.

$sel:format:ListPackageVersionAssets', listPackageVersionAssets_format - The format of the package that contains the returned package version assets. The valid package types are:

npm: A Node Package Manager (npm) package.

pypi: A Python Package Index (PyPI) package.

maven: A Maven package that contains compiled code in a distributable format, such as a JAR file.

$sel:package:ListPackageVersionAssets', listPackageVersionAssets_package - The name of the package that contains the returned package version assets.

$sel:packageVersion:ListPackageVersionAssets', listPackageVersionAssets_packageVersion - A string that contains the package version (for example, 3.5.2).

Request Lenses

listPackageVersionAssets_namespace :: Lens' ListPackageVersionAssets (Maybe Text)

The namespace of the package. The package component that specifies its namespace depends on its type. For example:

The namespace of a Maven package is its groupId.

The namespace of an npm package is its scope.

A Python package does not contain a corresponding component, so Python packages do not have a namespace.

listPackageVersionAssets_domainOwner :: Lens' ListPackageVersionAssets (Maybe Text)

The 12-digit account number of the AWS account that owns the domain. It does not include dashes or spaces.

listPackageVersionAssets_nextToken :: Lens' ListPackageVersionAssets (Maybe Text)

The token for the next set of results. Use the value returned in the previous response in the next request to retrieve the next set of results.

listPackageVersionAssets_maxResults :: Lens' ListPackageVersionAssets (Maybe Natural)

The maximum number of results to return per page.

listPackageVersionAssets_domain :: Lens' ListPackageVersionAssets Text

The name of the domain that contains the repository associated with the package version assets.

listPackageVersionAssets_repository :: Lens' ListPackageVersionAssets Text

The name of the repository that contains the package that contains the returned package version assets.

listPackageVersionAssets_format :: Lens' ListPackageVersionAssets PackageFormat

The format of the package that contains the returned package version assets. The valid package types are:

npm: A Node Package Manager (npm) package.

pypi: A Python Package Index (PyPI) package.

maven: A Maven package that contains compiled code in a distributable format, such as a JAR file.

listPackageVersionAssets_package :: Lens' ListPackageVersionAssets Text

The name of the package that contains the returned package version assets.

listPackageVersionAssets_packageVersion :: Lens' ListPackageVersionAssets Text

A string that contains the package version (for example, 3.5.2).

Destructuring the Response

data ListPackageVersionAssetsResponse

See: newListPackageVersionAssetsResponse smart constructor.

Constructors

ListPackageVersionAssetsResponse'

Fields

format :: Maybe PackageFormat

The format of the package that contains the returned package version assets.

namespace :: Maybe Text

The namespace of the package. The package component that specifies its namespace depends on its type. For example:

The namespace of a Maven package is its groupId.

The namespace of an npm package is its scope.

A Python package does not contain a corresponding component, so Python packages do not have a namespace.

versionRevision :: Maybe Text

The current revision associated with the package version.

nextToken :: Maybe Text

If there are additional results, this is the token for the next set of results.

version :: Maybe Text

The version of the package associated with the returned assets.

package :: Maybe Text

The name of the package that contains the returned package version assets.

assets :: Maybe [AssetSummary]

The returned list of AssetSummary objects.

httpStatus :: Int

The response's http status code.

newListPackageVersionAssetsResponse

Arguments

:: Int
$sel:httpStatus:ListPackageVersionAssetsResponse'

-> ListPackageVersionAssetsResponse

Create a value of ListPackageVersionAssetsResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:format:ListPackageVersionAssets', listPackageVersionAssetsResponse_format - The format of the package that contains the returned package version assets.

$sel:namespace:ListPackageVersionAssets', listPackageVersionAssetsResponse_namespace - The namespace of the package. The package component that specifies its namespace depends on its type. For example:

The namespace of a Maven package is its groupId.

The namespace of an npm package is its scope.

A Python package does not contain a corresponding component, so Python packages do not have a namespace.

$sel:versionRevision:ListPackageVersionAssetsResponse', listPackageVersionAssetsResponse_versionRevision - The current revision associated with the package version.

$sel:nextToken:ListPackageVersionAssets', listPackageVersionAssetsResponse_nextToken - If there are additional results, this is the token for the next set of results.

$sel:version:ListPackageVersionAssetsResponse', listPackageVersionAssetsResponse_version - The version of the package associated with the returned assets.

$sel:package:ListPackageVersionAssets', listPackageVersionAssetsResponse_package - The name of the package that contains the returned package version assets.

$sel:assets:ListPackageVersionAssetsResponse', listPackageVersionAssetsResponse_assets - The returned list of AssetSummary objects.

$sel:httpStatus:ListPackageVersionAssetsResponse', listPackageVersionAssetsResponse_httpStatus - The response's http status code.

Response Lenses

listPackageVersionAssetsResponse_format :: Lens' ListPackageVersionAssetsResponse (Maybe PackageFormat)

The format of the package that contains the returned package version assets.

listPackageVersionAssetsResponse_namespace :: Lens' ListPackageVersionAssetsResponse (Maybe Text)

The namespace of the package. The package component that specifies its namespace depends on its type. For example:

The namespace of a Maven package is its groupId.

The namespace of an npm package is its scope.

A Python package does not contain a corresponding component, so Python packages do not have a namespace.

listPackageVersionAssetsResponse_versionRevision :: Lens' ListPackageVersionAssetsResponse (Maybe Text)

The current revision associated with the package version.

listPackageVersionAssetsResponse_nextToken :: Lens' ListPackageVersionAssetsResponse (Maybe Text)

If there are additional results, this is the token for the next set of results.

listPackageVersionAssetsResponse_version :: Lens' ListPackageVersionAssetsResponse (Maybe Text)

The version of the package associated with the returned assets.

listPackageVersionAssetsResponse_package :: Lens' ListPackageVersionAssetsResponse (Maybe Text)

The name of the package that contains the returned package version assets.

listPackageVersionAssetsResponse_assets :: Lens' ListPackageVersionAssetsResponse (Maybe [AssetSummary])

The returned list of AssetSummary objects.

listPackageVersionAssetsResponse_httpStatus :: Lens' ListPackageVersionAssetsResponse Int

The response's http status code.