Amazonka.CodeArtifact.ListPackages

Description

Returns a list of PackageSummary objects for packages in a repository that match the request parameters.

This operation returns paginated results.

Creating a Request

data ListPackages

See: newListPackages smart constructor.

Constructors

ListPackages'

Fields

format :: Maybe PackageFormat

The format of the packages. The valid package types are:

npm: A Node Package Manager (npm) package.

pypi: A Python Package Index (PyPI) package.

maven: A Maven package that contains compiled code in a distributable format, such as a JAR file.

namespace :: Maybe Text

The namespace of the package. The package component that specifies its namespace depends on its type. For example:

The namespace of a Maven package is its groupId.

The namespace of an npm package is its scope.

A Python package does not contain a corresponding component, so Python packages do not have a namespace.

domainOwner :: Maybe Text

The 12-digit account number of the AWS account that owns the domain. It does not include dashes or spaces.

nextToken :: Maybe Text

The token for the next set of results. Use the value returned in the previous response in the next request to retrieve the next set of results.

packagePrefix :: Maybe Text

A prefix used to filter returned packages. Only packages with names that start with packagePrefix are returned.

maxResults :: Maybe Natural

The maximum number of results to return per page.

domain :: Text

The name of the domain that contains the repository that contains the requested list of packages.

repository :: Text

The name of the repository from which packages are to be listed.

newListPackages

Arguments

:: Text
$sel:domain:ListPackages'

-> Text
$sel:repository:ListPackages'

-> ListPackages

Create a value of ListPackages with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:format:ListPackages', listPackages_format - The format of the packages. The valid package types are:

npm: A Node Package Manager (npm) package.

pypi: A Python Package Index (PyPI) package.

maven: A Maven package that contains compiled code in a distributable format, such as a JAR file.

$sel:namespace:ListPackages', listPackages_namespace - The namespace of the package. The package component that specifies its namespace depends on its type. For example:

The namespace of a Maven package is its groupId.

The namespace of an npm package is its scope.

A Python package does not contain a corresponding component, so Python packages do not have a namespace.

$sel:domainOwner:ListPackages', listPackages_domainOwner - The 12-digit account number of the AWS account that owns the domain. It does not include dashes or spaces.

$sel:nextToken:ListPackages', listPackages_nextToken - The token for the next set of results. Use the value returned in the previous response in the next request to retrieve the next set of results.

$sel:packagePrefix:ListPackages', listPackages_packagePrefix - A prefix used to filter returned packages. Only packages with names that start with packagePrefix are returned.

$sel:maxResults:ListPackages', listPackages_maxResults - The maximum number of results to return per page.

$sel:domain:ListPackages', listPackages_domain - The name of the domain that contains the repository that contains the requested list of packages.

$sel:repository:ListPackages', listPackages_repository - The name of the repository from which packages are to be listed.

Request Lenses

listPackages_format :: Lens' ListPackages (Maybe PackageFormat)

The format of the packages. The valid package types are:

npm: A Node Package Manager (npm) package.

pypi: A Python Package Index (PyPI) package.

maven: A Maven package that contains compiled code in a distributable format, such as a JAR file.

listPackages_namespace :: Lens' ListPackages (Maybe Text)

The namespace of the package. The package component that specifies its namespace depends on its type. For example:

The namespace of a Maven package is its groupId.

The namespace of an npm package is its scope.

A Python package does not contain a corresponding component, so Python packages do not have a namespace.

listPackages_domainOwner :: Lens' ListPackages (Maybe Text)

The 12-digit account number of the AWS account that owns the domain. It does not include dashes or spaces.

listPackages_nextToken :: Lens' ListPackages (Maybe Text)

The token for the next set of results. Use the value returned in the previous response in the next request to retrieve the next set of results.

listPackages_packagePrefix :: Lens' ListPackages (Maybe Text)

A prefix used to filter returned packages. Only packages with names that start with packagePrefix are returned.

listPackages_maxResults :: Lens' ListPackages (Maybe Natural)

The maximum number of results to return per page.

listPackages_domain :: Lens' ListPackages Text

The name of the domain that contains the repository that contains the requested list of packages.

listPackages_repository :: Lens' ListPackages Text

The name of the repository from which packages are to be listed.

Destructuring the Response

data ListPackagesResponse

See: newListPackagesResponse smart constructor.

Constructors

ListPackagesResponse'

Fields

packages :: Maybe [PackageSummary]

The list of returned PackageSummary objects.

nextToken :: Maybe Text

If there are additional results, this is the token for the next set of results.

httpStatus :: Int

The response's http status code.

newListPackagesResponse

Arguments

:: Int
$sel:httpStatus:ListPackagesResponse'

-> ListPackagesResponse

Create a value of ListPackagesResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:packages:ListPackagesResponse', listPackagesResponse_packages - The list of returned PackageSummary objects.

$sel:nextToken:ListPackages', listPackagesResponse_nextToken - If there are additional results, this is the token for the next set of results.

$sel:httpStatus:ListPackagesResponse', listPackagesResponse_httpStatus - The response's http status code.

Response Lenses

listPackagesResponse_packages :: Lens' ListPackagesResponse (Maybe [PackageSummary])

The list of returned PackageSummary objects.

listPackagesResponse_nextToken :: Lens' ListPackagesResponse (Maybe Text)

If there are additional results, this is the token for the next set of results.

listPackagesResponse_httpStatus :: Lens' ListPackagesResponse Int

The response's http status code.