Amazonka.CodeArtifact.Types.DomainSummary

Description

Documentation

data DomainSummary

Information about a domain, including its name, Amazon Resource Name (ARN), and status. The ListDomains operation returns a list of DomainSummary objects.

See: newDomainSummary smart constructor.

Constructors

DomainSummary'

Fields

status :: Maybe DomainStatus

A string that contains the status of the domain. The valid values are:

Active

Deleted

arn :: Maybe Text

The ARN of the domain.

createdTime :: Maybe POSIX

A timestamp that contains the date and time the domain was created.

owner :: Maybe Text

The 12-digit account number of the AWS account that owns the domain. It does not include dashes or spaces.

name :: Maybe Text

The name of the domain.

encryptionKey :: Maybe Text

The key used to encrypt the domain.

newDomainSummary :: DomainSummary

Create a value of DomainSummary with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:status:DomainSummary', domainSummary_status - A string that contains the status of the domain. The valid values are:

Active

Deleted

$sel:arn:DomainSummary', domainSummary_arn - The ARN of the domain.

$sel:createdTime:DomainSummary', domainSummary_createdTime - A timestamp that contains the date and time the domain was created.

$sel:owner:DomainSummary', domainSummary_owner - The 12-digit account number of the AWS account that owns the domain. It does not include dashes or spaces.

$sel:name:DomainSummary', domainSummary_name - The name of the domain.

$sel:encryptionKey:DomainSummary', domainSummary_encryptionKey - The key used to encrypt the domain.

domainSummary_status :: Lens' DomainSummary (Maybe DomainStatus)

A string that contains the status of the domain. The valid values are:

Active

Deleted

domainSummary_arn :: Lens' DomainSummary (Maybe Text)

The ARN of the domain.

domainSummary_createdTime :: Lens' DomainSummary (Maybe UTCTime)

A timestamp that contains the date and time the domain was created.

domainSummary_owner :: Lens' DomainSummary (Maybe Text)

The 12-digit account number of the AWS account that owns the domain. It does not include dashes or spaces.

domainSummary_name :: Lens' DomainSummary (Maybe Text)

The name of the domain.

domainSummary_encryptionKey :: Lens' DomainSummary (Maybe Text)

The key used to encrypt the domain.