Amazonka.CodeArtifact.Types.PackageDependency

Description

Documentation

data PackageDependency

Details about a package dependency.

See: newPackageDependency smart constructor.

Constructors

PackageDependency'

Fields

dependencyType :: Maybe Text

The type of a package dependency. The possible values depend on the package type. Example types are compile, runtime, and test for Maven packages, and dev, prod, and optional for npm packages.

namespace :: Maybe Text

The namespace of the package. The package component that specifies its namespace depends on its type. For example:

The namespace of a Maven package is its groupId.

The namespace of an npm package is its scope.

A Python package does not contain a corresponding component, so Python packages do not have a namespace.

package :: Maybe Text

The name of the package that this package depends on.

versionRequirement :: Maybe Text

The required version, or version range, of the package that this package depends on. The version format is specific to the package type. For example, the following are possible valid required versions: 1.2.3, ^2.3.4, or 4.x.

newPackageDependency :: PackageDependency

Create a value of PackageDependency with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:dependencyType:PackageDependency', packageDependency_dependencyType - The type of a package dependency. The possible values depend on the package type. Example types are compile, runtime, and test for Maven packages, and dev, prod, and optional for npm packages.

$sel:namespace:PackageDependency', packageDependency_namespace - The namespace of the package. The package component that specifies its namespace depends on its type. For example:

The namespace of a Maven package is its groupId.

The namespace of an npm package is its scope.

A Python package does not contain a corresponding component, so Python packages do not have a namespace.

$sel:package:PackageDependency', packageDependency_package - The name of the package that this package depends on.

$sel:versionRequirement:PackageDependency', packageDependency_versionRequirement - The required version, or version range, of the package that this package depends on. The version format is specific to the package type. For example, the following are possible valid required versions: 1.2.3, ^2.3.4, or 4.x.

packageDependency_dependencyType :: Lens' PackageDependency (Maybe Text)

The type of a package dependency. The possible values depend on the package type. Example types are compile, runtime, and test for Maven packages, and dev, prod, and optional for npm packages.

packageDependency_namespace :: Lens' PackageDependency (Maybe Text)

The namespace of the package. The package component that specifies its namespace depends on its type. For example:

The namespace of a Maven package is its groupId.

The namespace of an npm package is its scope.

A Python package does not contain a corresponding component, so Python packages do not have a namespace.

packageDependency_package :: Lens' PackageDependency (Maybe Text)

The name of the package that this package depends on.

packageDependency_versionRequirement :: Lens' PackageDependency (Maybe Text)

The required version, or version range, of the package that this package depends on. The version format is specific to the package type. For example, the following are possible valid required versions: 1.2.3, ^2.3.4, or 4.x.