Amazonka.CodeArtifact.Types.PackageVersionDescription

Description

Documentation

data PackageVersionDescription

Details about a package version.

See: newPackageVersionDescription smart constructor.

Constructors

PackageVersionDescription'

Fields

summary :: Maybe Text

A summary of the package version. The summary is extracted from the package. The information in and detail level of the summary depends on the package version's format.

homePage :: Maybe Text

The homepage associated with the package.

status :: Maybe PackageVersionStatus

A string that contains the status of the package version. It can be one of the following:

Published

Unfinished

Unlisted

Archived

Disposed

sourceCodeRepository :: Maybe Text

The repository for the source code in the package version, or the source code used to build it.

packageName :: Maybe Text

The name of the requested package.

publishedTime :: Maybe POSIX

A timestamp that contains the date and time the package version was published.

format :: Maybe PackageFormat

The format of the package version. The valid package formats are:

npm: A Node Package Manager (npm) package.

pypi: A Python Package Index (PyPI) package.

maven: A Maven package that contains compiled code in a distributable format, such as a JAR file.

namespace :: Maybe Text

The namespace of the package. The package component that specifies its namespace depends on its type. For example:

The namespace of a Maven package is its groupId.

The namespace of an npm package is its scope.

A Python package does not contain a corresponding component, so Python packages do not have a namespace.

version :: Maybe Text

The version of the package.

displayName :: Maybe Text

The name of the package that is displayed. The displayName varies depending on the package version's format. For example, if an npm package is named ui, is in the namespace vue, and has the format npm, then the displayName is @vue/ui.

licenses :: Maybe [LicenseInfo]

Information about licenses associated with the package version.

revision :: Maybe Text

The revision of the package version.

newPackageVersionDescription :: PackageVersionDescription

Create a value of PackageVersionDescription with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:summary:PackageVersionDescription', packageVersionDescription_summary - A summary of the package version. The summary is extracted from the package. The information in and detail level of the summary depends on the package version's format.

$sel:homePage:PackageVersionDescription', packageVersionDescription_homePage - The homepage associated with the package.

$sel:status:PackageVersionDescription', packageVersionDescription_status - A string that contains the status of the package version. It can be one of the following:

Published

Unfinished

Unlisted

Archived

Disposed

$sel:sourceCodeRepository:PackageVersionDescription', packageVersionDescription_sourceCodeRepository - The repository for the source code in the package version, or the source code used to build it.

$sel:packageName:PackageVersionDescription', packageVersionDescription_packageName - The name of the requested package.

$sel:publishedTime:PackageVersionDescription', packageVersionDescription_publishedTime - A timestamp that contains the date and time the package version was published.

$sel:format:PackageVersionDescription', packageVersionDescription_format - The format of the package version. The valid package formats are:

npm: A Node Package Manager (npm) package.

pypi: A Python Package Index (PyPI) package.

maven: A Maven package that contains compiled code in a distributable format, such as a JAR file.

$sel:namespace:PackageVersionDescription', packageVersionDescription_namespace - The namespace of the package. The package component that specifies its namespace depends on its type. For example:

The namespace of a Maven package is its groupId.

The namespace of an npm package is its scope.

A Python package does not contain a corresponding component, so Python packages do not have a namespace.

$sel:version:PackageVersionDescription', packageVersionDescription_version - The version of the package.

$sel:displayName:PackageVersionDescription', packageVersionDescription_displayName - The name of the package that is displayed. The displayName varies depending on the package version's format. For example, if an npm package is named ui, is in the namespace vue, and has the format npm, then the displayName is @vue/ui.

$sel:licenses:PackageVersionDescription', packageVersionDescription_licenses - Information about licenses associated with the package version.

$sel:revision:PackageVersionDescription', packageVersionDescription_revision - The revision of the package version.

packageVersionDescription_summary :: Lens' PackageVersionDescription (Maybe Text)

A summary of the package version. The summary is extracted from the package. The information in and detail level of the summary depends on the package version's format.

packageVersionDescription_homePage :: Lens' PackageVersionDescription (Maybe Text)

The homepage associated with the package.

packageVersionDescription_status :: Lens' PackageVersionDescription (Maybe PackageVersionStatus)

A string that contains the status of the package version. It can be one of the following:

Published

Unfinished

Unlisted

Archived

Disposed

packageVersionDescription_sourceCodeRepository :: Lens' PackageVersionDescription (Maybe Text)

The repository for the source code in the package version, or the source code used to build it.

packageVersionDescription_packageName :: Lens' PackageVersionDescription (Maybe Text)

The name of the requested package.

packageVersionDescription_publishedTime :: Lens' PackageVersionDescription (Maybe UTCTime)

A timestamp that contains the date and time the package version was published.

packageVersionDescription_format :: Lens' PackageVersionDescription (Maybe PackageFormat)

The format of the package version. The valid package formats are:

npm: A Node Package Manager (npm) package.

pypi: A Python Package Index (PyPI) package.

maven: A Maven package that contains compiled code in a distributable format, such as a JAR file.

packageVersionDescription_namespace :: Lens' PackageVersionDescription (Maybe Text)

The namespace of the package. The package component that specifies its namespace depends on its type. For example:

The namespace of a Maven package is its groupId.

The namespace of an npm package is its scope.

A Python package does not contain a corresponding component, so Python packages do not have a namespace.

packageVersionDescription_version :: Lens' PackageVersionDescription (Maybe Text)

The version of the package.

packageVersionDescription_displayName :: Lens' PackageVersionDescription (Maybe Text)

The name of the package that is displayed. The displayName varies depending on the package version's format. For example, if an npm package is named ui, is in the namespace vue, and has the format npm, then the displayName is @vue/ui.

packageVersionDescription_licenses :: Lens' PackageVersionDescription (Maybe [LicenseInfo])

Information about licenses associated with the package version.

packageVersionDescription_revision :: Lens' PackageVersionDescription (Maybe Text)

The revision of the package version.