Amazonka.CodeBuild.ListBuildBatchesForProject

Description

Retrieves the identifiers of the build batches for a specific project.

This operation returns paginated results.

Creating a Request

data ListBuildBatchesForProject

See: newListBuildBatchesForProject smart constructor.

Constructors

ListBuildBatchesForProject'

Fields

sortOrder :: Maybe SortOrderType

Specifies the sort order of the returned items. Valid values include:

ASCENDING: List the batch build identifiers in ascending order by identifier.

DESCENDING: List the batch build identifiers in descending order by identifier.

nextToken :: Maybe Text

The nextToken value returned from a previous call to ListBuildBatchesForProject. This specifies the next item to return. To return the beginning of the list, exclude this parameter.

projectName :: Maybe Text

The name of the project.

filter' :: Maybe BuildBatchFilter

A BuildBatchFilter object that specifies the filters for the search.

maxResults :: Maybe Natural

The maximum number of results to return.

newListBuildBatchesForProject :: ListBuildBatchesForProject

Create a value of ListBuildBatchesForProject with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:sortOrder:ListBuildBatchesForProject', listBuildBatchesForProject_sortOrder - Specifies the sort order of the returned items. Valid values include:

ASCENDING: List the batch build identifiers in ascending order by identifier.

DESCENDING: List the batch build identifiers in descending order by identifier.

$sel:nextToken:ListBuildBatchesForProject', listBuildBatchesForProject_nextToken - The nextToken value returned from a previous call to ListBuildBatchesForProject. This specifies the next item to return. To return the beginning of the list, exclude this parameter.

$sel:projectName:ListBuildBatchesForProject', listBuildBatchesForProject_projectName - The name of the project.

$sel:filter':ListBuildBatchesForProject', listBuildBatchesForProject_filter - A BuildBatchFilter object that specifies the filters for the search.

$sel:maxResults:ListBuildBatchesForProject', listBuildBatchesForProject_maxResults - The maximum number of results to return.

Request Lenses

listBuildBatchesForProject_sortOrder :: Lens' ListBuildBatchesForProject (Maybe SortOrderType)

Specifies the sort order of the returned items. Valid values include:

ASCENDING: List the batch build identifiers in ascending order by identifier.

DESCENDING: List the batch build identifiers in descending order by identifier.

listBuildBatchesForProject_nextToken :: Lens' ListBuildBatchesForProject (Maybe Text)

The nextToken value returned from a previous call to ListBuildBatchesForProject. This specifies the next item to return. To return the beginning of the list, exclude this parameter.

listBuildBatchesForProject_projectName :: Lens' ListBuildBatchesForProject (Maybe Text)

The name of the project.

listBuildBatchesForProject_filter :: Lens' ListBuildBatchesForProject (Maybe BuildBatchFilter)

A BuildBatchFilter object that specifies the filters for the search.

listBuildBatchesForProject_maxResults :: Lens' ListBuildBatchesForProject (Maybe Natural)

The maximum number of results to return.

Destructuring the Response

data ListBuildBatchesForProjectResponse

See: newListBuildBatchesForProjectResponse smart constructor.

Constructors

ListBuildBatchesForProjectResponse'

Fields

ids :: Maybe [Text]

An array of strings that contains the batch build identifiers.

nextToken :: Maybe Text

If there are more items to return, this contains a token that is passed to a subsequent call to ListBuildBatchesForProject to retrieve the next set of items.

httpStatus :: Int

The response's http status code.

newListBuildBatchesForProjectResponse

Arguments

:: Int
$sel:httpStatus:ListBuildBatchesForProjectResponse'

-> ListBuildBatchesForProjectResponse

Create a value of ListBuildBatchesForProjectResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:ids:ListBuildBatchesForProjectResponse', listBuildBatchesForProjectResponse_ids - An array of strings that contains the batch build identifiers.

$sel:nextToken:ListBuildBatchesForProject', listBuildBatchesForProjectResponse_nextToken - If there are more items to return, this contains a token that is passed to a subsequent call to ListBuildBatchesForProject to retrieve the next set of items.

$sel:httpStatus:ListBuildBatchesForProjectResponse', listBuildBatchesForProjectResponse_httpStatus - The response's http status code.

Response Lenses

listBuildBatchesForProjectResponse_ids :: Lens' ListBuildBatchesForProjectResponse (Maybe [Text])

An array of strings that contains the batch build identifiers.

listBuildBatchesForProjectResponse_nextToken :: Lens' ListBuildBatchesForProjectResponse (Maybe Text)

If there are more items to return, this contains a token that is passed to a subsequent call to ListBuildBatchesForProject to retrieve the next set of items.

listBuildBatchesForProjectResponse_httpStatus :: Lens' ListBuildBatchesForProjectResponse Int

The response's http status code.