Amazonka.CodeBuild.RetryBuildBatch

Description

Restarts a failed batch build. Only batch builds that have failed can be retried.

Creating a Request

data RetryBuildBatch

See: newRetryBuildBatch smart constructor.

Constructors

RetryBuildBatch'

Fields

idempotencyToken :: Maybe Text

A unique, case sensitive identifier you provide to ensure the idempotency of the RetryBuildBatch request. The token is included in the RetryBuildBatch request and is valid for five minutes. If you repeat the RetryBuildBatch request with the same token, but change a parameter, CodeBuild returns a parameter mismatch error.

id :: Maybe Text

Specifies the identifier of the batch build to restart.

retryType :: Maybe RetryBuildBatchType

Specifies the type of retry to perform.

newRetryBuildBatch :: RetryBuildBatch

Create a value of RetryBuildBatch with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:idempotencyToken:RetryBuildBatch', retryBuildBatch_idempotencyToken - A unique, case sensitive identifier you provide to ensure the idempotency of the RetryBuildBatch request. The token is included in the RetryBuildBatch request and is valid for five minutes. If you repeat the RetryBuildBatch request with the same token, but change a parameter, CodeBuild returns a parameter mismatch error.

$sel:id:RetryBuildBatch', retryBuildBatch_id - Specifies the identifier of the batch build to restart.

$sel:retryType:RetryBuildBatch', retryBuildBatch_retryType - Specifies the type of retry to perform.

Request Lenses

retryBuildBatch_idempotencyToken :: Lens' RetryBuildBatch (Maybe Text)

A unique, case sensitive identifier you provide to ensure the idempotency of the RetryBuildBatch request. The token is included in the RetryBuildBatch request and is valid for five minutes. If you repeat the RetryBuildBatch request with the same token, but change a parameter, CodeBuild returns a parameter mismatch error.

retryBuildBatch_id :: Lens' RetryBuildBatch (Maybe Text)

Specifies the identifier of the batch build to restart.

retryBuildBatch_retryType :: Lens' RetryBuildBatch (Maybe RetryBuildBatchType)

Specifies the type of retry to perform.

Destructuring the Response

data RetryBuildBatchResponse

See: newRetryBuildBatchResponse smart constructor.

Constructors

RetryBuildBatchResponse'

Fields

buildBatch :: Maybe BuildBatch

httpStatus :: Int

The response's http status code.

newRetryBuildBatchResponse

Arguments

:: Int
$sel:httpStatus:RetryBuildBatchResponse'

-> RetryBuildBatchResponse

Create a value of RetryBuildBatchResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:buildBatch:RetryBuildBatchResponse', retryBuildBatchResponse_buildBatch - Undocumented member.

$sel:httpStatus:RetryBuildBatchResponse', retryBuildBatchResponse_httpStatus - The response's http status code.

Response Lenses

retryBuildBatchResponse_buildBatch :: Lens' RetryBuildBatchResponse (Maybe BuildBatch)

Undocumented member.

retryBuildBatchResponse_httpStatus :: Lens' RetryBuildBatchResponse Int

The response's http status code.