Amazonka.CodeBuild.StartBuild

Description

Starts running a build.

Creating a Request

data StartBuild

See: newStartBuild smart constructor.

Constructors

StartBuild'

Fields

encryptionKeyOverride :: Maybe Text

The Key Management Service customer master key (CMK) that overrides the one specified in the build project. The CMK key encrypts the build output artifacts.

You can use a cross-account KMS key to encrypt the build output artifacts if your service role has permission to that key.

You can specify either the Amazon Resource Name (ARN) of the CMK or, if available, the CMK's alias (using the format alias/<alias-name>).

sourceLocationOverride :: Maybe Text

A location that overrides, for this build, the source location for the one defined in the build project.

environmentVariablesOverride :: Maybe [EnvironmentVariable]

A set of environment variables that overrides, for this build only, the latest ones already defined in the build project.

buildStatusConfigOverride :: Maybe BuildStatusConfig

Contains information that defines how the build project reports the build status to the source provider. This option is only used when the source provider is GITHUB, GITHUB_ENTERPRISE, or BITBUCKET.

idempotencyToken :: Maybe Text

A unique, case sensitive identifier you provide to ensure the idempotency of the StartBuild request. The token is included in the StartBuild request and is valid for 5 minutes. If you repeat the StartBuild request with the same token, but change a parameter, CodeBuild returns a parameter mismatch error.

debugSessionEnabled :: Maybe Bool

Specifies if session debugging is enabled for this build. For more information, see Viewing a running build in Session Manager.

registryCredentialOverride :: Maybe RegistryCredential

The credentials for access to a private registry.

timeoutInMinutesOverride :: Maybe Natural

The number of build timeout minutes, from 5 to 480 (8 hours), that overrides, for this build only, the latest setting already defined in the build project.

serviceRoleOverride :: Maybe Text

The name of a service role for this build that overrides the one specified in the build project.

cacheOverride :: Maybe ProjectCache

A ProjectCache object specified for this build that overrides the one defined in the build project.

queuedTimeoutInMinutesOverride :: Maybe Natural

The number of minutes a build is allowed to be queued before it times out.

secondarySourcesOverride :: Maybe [ProjectSource]

An array of ProjectSource objects.

gitCloneDepthOverride :: Maybe Natural

The user-defined depth of history, with a minimum value of 0, that overrides, for this build only, any previous depth of history defined in the build project.

imagePullCredentialsTypeOverride :: Maybe ImagePullCredentialsType

The type of credentials CodeBuild uses to pull images in your build. There are two valid values:

CODEBUILD

Specifies that CodeBuild uses its own credentials. This requires that you modify your ECR repository policy to trust CodeBuild's service principal.

SERVICE_ROLE

Specifies that CodeBuild uses your build project's service role.

When using a cross-account or private registry image, you must use SERVICE_ROLE credentials. When using an CodeBuild curated image, you must use CODEBUILD credentials.

logsConfigOverride :: Maybe LogsConfig

Log settings for this build that override the log settings defined in the build project.

sourceAuthOverride :: Maybe SourceAuth

An authorization type for this build that overrides the one defined in the build project. This override applies only if the build project's source is BitBucket or GitHub.

gitSubmodulesConfigOverride :: Maybe GitSubmodulesConfig

Information about the Git submodules configuration for this build of an CodeBuild build project.

environmentTypeOverride :: Maybe EnvironmentType

A container type for this build that overrides the one specified in the build project.

certificateOverride :: Maybe Text

The name of a certificate for this build that overrides the one specified in the build project.

computeTypeOverride :: Maybe ComputeType

The name of a compute type for this build that overrides the one specified in the build project.

privilegedModeOverride :: Maybe Bool

Enable this flag to override privileged mode in the build project.

sourceVersion :: Maybe Text

The version of the build input to be built, for this build only. If not specified, the latest version is used. If specified, the contents depends on the source provider:

CodeCommit

The commit ID, branch, or Git tag to use.

GitHub

The commit ID, pull request ID, branch name, or tag name that corresponds to the version of the source code you want to build. If a pull request ID is specified, it must use the format pr/pull-request-ID (for example pr/25). If a branch name is specified, the branch's HEAD commit ID is used. If not specified, the default branch's HEAD commit ID is used.

Bitbucket

The commit ID, branch name, or tag name that corresponds to the version of the source code you want to build. If a branch name is specified, the branch's HEAD commit ID is used. If not specified, the default branch's HEAD commit ID is used.

Amazon S3

The version ID of the object that represents the build input ZIP file to use.

If sourceVersion is specified at the project level, then this sourceVersion (at the build level) takes precedence.

For more information, see Source Version Sample with CodeBuild in the CodeBuild User Guide.

buildspecOverride :: Maybe Text

A buildspec file declaration that overrides, for this build only, the latest one already defined in the build project.

If this value is set, it can be either an inline buildspec definition, the path to an alternate buildspec file relative to the value of the built-in CODEBUILD_SRC_DIR environment variable, or the path to an S3 bucket. The bucket must be in the same Amazon Web Services Region as the build project. Specify the buildspec file using its ARN (for example, arn:aws:s3:::my-codebuild-sample2/buildspec.yml). If this value is not provided or is set to an empty string, the source code must contain a buildspec file in its root directory. For more information, see Buildspec File Name and Storage Location.

secondarySourcesVersionOverride :: Maybe [ProjectSourceVersion]

An array of ProjectSourceVersion objects that specify one or more versions of the project's secondary sources to be used for this build only.

reportBuildStatusOverride :: Maybe Bool

Set to true to report to your source provider the status of a build's start and completion. If you use this option with a source provider other than GitHub, GitHub Enterprise, or Bitbucket, an invalidInputException is thrown.

To be able to report the build status to the source provider, the user associated with the source provider must have write access to the repo. If the user does not have write access, the build status cannot be updated. For more information, see Source provider access in the CodeBuild User Guide.

The status of a build triggered by a webhook is always reported to your source provider.

insecureSslOverride :: Maybe Bool

Enable this flag to override the insecure SSL setting that is specified in the build project. The insecure SSL setting determines whether to ignore SSL warnings while connecting to the project source code. This override applies only if the build's source is GitHub Enterprise.

imageOverride :: Maybe Text

The name of an image for this build that overrides the one specified in the build project.

secondaryArtifactsOverride :: Maybe [ProjectArtifacts]

An array of ProjectArtifacts objects.

artifactsOverride :: Maybe ProjectArtifacts

Build output artifact settings that override, for this build only, the latest ones already defined in the build project.

sourceTypeOverride :: Maybe SourceType

A source input type, for this build, that overrides the source input defined in the build project.

projectName :: Text

The name of the CodeBuild build project to start running a build.

newStartBuild

Arguments

:: Text
$sel:projectName:StartBuild'

-> StartBuild

Create a value of StartBuild with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:encryptionKeyOverride:StartBuild', startBuild_encryptionKeyOverride - The Key Management Service customer master key (CMK) that overrides the one specified in the build project. The CMK key encrypts the build output artifacts.

You can use a cross-account KMS key to encrypt the build output artifacts if your service role has permission to that key.

You can specify either the Amazon Resource Name (ARN) of the CMK or, if available, the CMK's alias (using the format alias/<alias-name>).

$sel:sourceLocationOverride:StartBuild', startBuild_sourceLocationOverride - A location that overrides, for this build, the source location for the one defined in the build project.

$sel:environmentVariablesOverride:StartBuild', startBuild_environmentVariablesOverride - A set of environment variables that overrides, for this build only, the latest ones already defined in the build project.

$sel:buildStatusConfigOverride:StartBuild', startBuild_buildStatusConfigOverride - Contains information that defines how the build project reports the build status to the source provider. This option is only used when the source provider is GITHUB, GITHUB_ENTERPRISE, or BITBUCKET.

$sel:idempotencyToken:StartBuild', startBuild_idempotencyToken - A unique, case sensitive identifier you provide to ensure the idempotency of the StartBuild request. The token is included in the StartBuild request and is valid for 5 minutes. If you repeat the StartBuild request with the same token, but change a parameter, CodeBuild returns a parameter mismatch error.

$sel:debugSessionEnabled:StartBuild', startBuild_debugSessionEnabled - Specifies if session debugging is enabled for this build. For more information, see Viewing a running build in Session Manager.

$sel:registryCredentialOverride:StartBuild', startBuild_registryCredentialOverride - The credentials for access to a private registry.

$sel:timeoutInMinutesOverride:StartBuild', startBuild_timeoutInMinutesOverride - The number of build timeout minutes, from 5 to 480 (8 hours), that overrides, for this build only, the latest setting already defined in the build project.

$sel:serviceRoleOverride:StartBuild', startBuild_serviceRoleOverride - The name of a service role for this build that overrides the one specified in the build project.

$sel:cacheOverride:StartBuild', startBuild_cacheOverride - A ProjectCache object specified for this build that overrides the one defined in the build project.

$sel:queuedTimeoutInMinutesOverride:StartBuild', startBuild_queuedTimeoutInMinutesOverride - The number of minutes a build is allowed to be queued before it times out.

$sel:secondarySourcesOverride:StartBuild', startBuild_secondarySourcesOverride - An array of ProjectSource objects.

$sel:gitCloneDepthOverride:StartBuild', startBuild_gitCloneDepthOverride - The user-defined depth of history, with a minimum value of 0, that overrides, for this build only, any previous depth of history defined in the build project.

$sel:imagePullCredentialsTypeOverride:StartBuild', startBuild_imagePullCredentialsTypeOverride - The type of credentials CodeBuild uses to pull images in your build. There are two valid values:

CODEBUILD

Specifies that CodeBuild uses its own credentials. This requires that you modify your ECR repository policy to trust CodeBuild's service principal.

SERVICE_ROLE

Specifies that CodeBuild uses your build project's service role.

When using a cross-account or private registry image, you must use SERVICE_ROLE credentials. When using an CodeBuild curated image, you must use CODEBUILD credentials.

$sel:logsConfigOverride:StartBuild', startBuild_logsConfigOverride - Log settings for this build that override the log settings defined in the build project.

$sel:sourceAuthOverride:StartBuild', startBuild_sourceAuthOverride - An authorization type for this build that overrides the one defined in the build project. This override applies only if the build project's source is BitBucket or GitHub.

$sel:gitSubmodulesConfigOverride:StartBuild', startBuild_gitSubmodulesConfigOverride - Information about the Git submodules configuration for this build of an CodeBuild build project.

$sel:environmentTypeOverride:StartBuild', startBuild_environmentTypeOverride - A container type for this build that overrides the one specified in the build project.

$sel:certificateOverride:StartBuild', startBuild_certificateOverride - The name of a certificate for this build that overrides the one specified in the build project.

$sel:computeTypeOverride:StartBuild', startBuild_computeTypeOverride - The name of a compute type for this build that overrides the one specified in the build project.

$sel:privilegedModeOverride:StartBuild', startBuild_privilegedModeOverride - Enable this flag to override privileged mode in the build project.

$sel:sourceVersion:StartBuild', startBuild_sourceVersion - The version of the build input to be built, for this build only. If not specified, the latest version is used. If specified, the contents depends on the source provider:

CodeCommit

The commit ID, branch, or Git tag to use.

GitHub

The commit ID, pull request ID, branch name, or tag name that corresponds to the version of the source code you want to build. If a pull request ID is specified, it must use the format pr/pull-request-ID (for example pr/25). If a branch name is specified, the branch's HEAD commit ID is used. If not specified, the default branch's HEAD commit ID is used.

Bitbucket

The commit ID, branch name, or tag name that corresponds to the version of the source code you want to build. If a branch name is specified, the branch's HEAD commit ID is used. If not specified, the default branch's HEAD commit ID is used.

Amazon S3

The version ID of the object that represents the build input ZIP file to use.

If sourceVersion is specified at the project level, then this sourceVersion (at the build level) takes precedence.

For more information, see Source Version Sample with CodeBuild in the CodeBuild User Guide.

$sel:buildspecOverride:StartBuild', startBuild_buildspecOverride - A buildspec file declaration that overrides, for this build only, the latest one already defined in the build project.

If this value is set, it can be either an inline buildspec definition, the path to an alternate buildspec file relative to the value of the built-in CODEBUILD_SRC_DIR environment variable, or the path to an S3 bucket. The bucket must be in the same Amazon Web Services Region as the build project. Specify the buildspec file using its ARN (for example, arn:aws:s3:::my-codebuild-sample2/buildspec.yml). If this value is not provided or is set to an empty string, the source code must contain a buildspec file in its root directory. For more information, see Buildspec File Name and Storage Location.

$sel:secondarySourcesVersionOverride:StartBuild', startBuild_secondarySourcesVersionOverride - An array of ProjectSourceVersion objects that specify one or more versions of the project's secondary sources to be used for this build only.

$sel:reportBuildStatusOverride:StartBuild', startBuild_reportBuildStatusOverride - Set to true to report to your source provider the status of a build's start and completion. If you use this option with a source provider other than GitHub, GitHub Enterprise, or Bitbucket, an invalidInputException is thrown.

To be able to report the build status to the source provider, the user associated with the source provider must have write access to the repo. If the user does not have write access, the build status cannot be updated. For more information, see Source provider access in the CodeBuild User Guide.

The status of a build triggered by a webhook is always reported to your source provider.

$sel:insecureSslOverride:StartBuild', startBuild_insecureSslOverride - Enable this flag to override the insecure SSL setting that is specified in the build project. The insecure SSL setting determines whether to ignore SSL warnings while connecting to the project source code. This override applies only if the build's source is GitHub Enterprise.

$sel:imageOverride:StartBuild', startBuild_imageOverride - The name of an image for this build that overrides the one specified in the build project.

$sel:secondaryArtifactsOverride:StartBuild', startBuild_secondaryArtifactsOverride - An array of ProjectArtifacts objects.

$sel:artifactsOverride:StartBuild', startBuild_artifactsOverride - Build output artifact settings that override, for this build only, the latest ones already defined in the build project.

$sel:sourceTypeOverride:StartBuild', startBuild_sourceTypeOverride - A source input type, for this build, that overrides the source input defined in the build project.

$sel:projectName:StartBuild', startBuild_projectName - The name of the CodeBuild build project to start running a build.

Request Lenses

startBuild_encryptionKeyOverride :: Lens' StartBuild (Maybe Text)

The Key Management Service customer master key (CMK) that overrides the one specified in the build project. The CMK key encrypts the build output artifacts.

You can use a cross-account KMS key to encrypt the build output artifacts if your service role has permission to that key.

You can specify either the Amazon Resource Name (ARN) of the CMK or, if available, the CMK's alias (using the format alias/<alias-name>).

startBuild_sourceLocationOverride :: Lens' StartBuild (Maybe Text)

A location that overrides, for this build, the source location for the one defined in the build project.

startBuild_environmentVariablesOverride :: Lens' StartBuild (Maybe [EnvironmentVariable])

A set of environment variables that overrides, for this build only, the latest ones already defined in the build project.

startBuild_buildStatusConfigOverride :: Lens' StartBuild (Maybe BuildStatusConfig)

Contains information that defines how the build project reports the build status to the source provider. This option is only used when the source provider is GITHUB, GITHUB_ENTERPRISE, or BITBUCKET.

startBuild_idempotencyToken :: Lens' StartBuild (Maybe Text)

A unique, case sensitive identifier you provide to ensure the idempotency of the StartBuild request. The token is included in the StartBuild request and is valid for 5 minutes. If you repeat the StartBuild request with the same token, but change a parameter, CodeBuild returns a parameter mismatch error.

startBuild_debugSessionEnabled :: Lens' StartBuild (Maybe Bool)

Specifies if session debugging is enabled for this build. For more information, see Viewing a running build in Session Manager.

startBuild_registryCredentialOverride :: Lens' StartBuild (Maybe RegistryCredential)

The credentials for access to a private registry.

startBuild_timeoutInMinutesOverride :: Lens' StartBuild (Maybe Natural)

The number of build timeout minutes, from 5 to 480 (8 hours), that overrides, for this build only, the latest setting already defined in the build project.

startBuild_serviceRoleOverride :: Lens' StartBuild (Maybe Text)

The name of a service role for this build that overrides the one specified in the build project.

startBuild_cacheOverride :: Lens' StartBuild (Maybe ProjectCache)

A ProjectCache object specified for this build that overrides the one defined in the build project.

startBuild_queuedTimeoutInMinutesOverride :: Lens' StartBuild (Maybe Natural)

The number of minutes a build is allowed to be queued before it times out.

startBuild_secondarySourcesOverride :: Lens' StartBuild (Maybe [ProjectSource])

An array of ProjectSource objects.

startBuild_gitCloneDepthOverride :: Lens' StartBuild (Maybe Natural)

The user-defined depth of history, with a minimum value of 0, that overrides, for this build only, any previous depth of history defined in the build project.

startBuild_imagePullCredentialsTypeOverride :: Lens' StartBuild (Maybe ImagePullCredentialsType)

The type of credentials CodeBuild uses to pull images in your build. There are two valid values:

CODEBUILD

Specifies that CodeBuild uses its own credentials. This requires that you modify your ECR repository policy to trust CodeBuild's service principal.

SERVICE_ROLE

Specifies that CodeBuild uses your build project's service role.

When using a cross-account or private registry image, you must use SERVICE_ROLE credentials. When using an CodeBuild curated image, you must use CODEBUILD credentials.

startBuild_logsConfigOverride :: Lens' StartBuild (Maybe LogsConfig)

Log settings for this build that override the log settings defined in the build project.

startBuild_sourceAuthOverride :: Lens' StartBuild (Maybe SourceAuth)

An authorization type for this build that overrides the one defined in the build project. This override applies only if the build project's source is BitBucket or GitHub.

startBuild_gitSubmodulesConfigOverride :: Lens' StartBuild (Maybe GitSubmodulesConfig)

Information about the Git submodules configuration for this build of an CodeBuild build project.

startBuild_environmentTypeOverride :: Lens' StartBuild (Maybe EnvironmentType)

A container type for this build that overrides the one specified in the build project.

startBuild_certificateOverride :: Lens' StartBuild (Maybe Text)

The name of a certificate for this build that overrides the one specified in the build project.

startBuild_computeTypeOverride :: Lens' StartBuild (Maybe ComputeType)

The name of a compute type for this build that overrides the one specified in the build project.

startBuild_privilegedModeOverride :: Lens' StartBuild (Maybe Bool)

Enable this flag to override privileged mode in the build project.

startBuild_sourceVersion :: Lens' StartBuild (Maybe Text)

The version of the build input to be built, for this build only. If not specified, the latest version is used. If specified, the contents depends on the source provider:

CodeCommit

The commit ID, branch, or Git tag to use.

GitHub

The commit ID, pull request ID, branch name, or tag name that corresponds to the version of the source code you want to build. If a pull request ID is specified, it must use the format pr/pull-request-ID (for example pr/25). If a branch name is specified, the branch's HEAD commit ID is used. If not specified, the default branch's HEAD commit ID is used.

Bitbucket

The commit ID, branch name, or tag name that corresponds to the version of the source code you want to build. If a branch name is specified, the branch's HEAD commit ID is used. If not specified, the default branch's HEAD commit ID is used.

Amazon S3

The version ID of the object that represents the build input ZIP file to use.

If sourceVersion is specified at the project level, then this sourceVersion (at the build level) takes precedence.

For more information, see Source Version Sample with CodeBuild in the CodeBuild User Guide.

startBuild_buildspecOverride :: Lens' StartBuild (Maybe Text)

A buildspec file declaration that overrides, for this build only, the latest one already defined in the build project.

If this value is set, it can be either an inline buildspec definition, the path to an alternate buildspec file relative to the value of the built-in CODEBUILD_SRC_DIR environment variable, or the path to an S3 bucket. The bucket must be in the same Amazon Web Services Region as the build project. Specify the buildspec file using its ARN (for example, arn:aws:s3:::my-codebuild-sample2/buildspec.yml). If this value is not provided or is set to an empty string, the source code must contain a buildspec file in its root directory. For more information, see Buildspec File Name and Storage Location.

startBuild_secondarySourcesVersionOverride :: Lens' StartBuild (Maybe [ProjectSourceVersion])

An array of ProjectSourceVersion objects that specify one or more versions of the project's secondary sources to be used for this build only.

startBuild_reportBuildStatusOverride :: Lens' StartBuild (Maybe Bool)

Set to true to report to your source provider the status of a build's start and completion. If you use this option with a source provider other than GitHub, GitHub Enterprise, or Bitbucket, an invalidInputException is thrown.

To be able to report the build status to the source provider, the user associated with the source provider must have write access to the repo. If the user does not have write access, the build status cannot be updated. For more information, see Source provider access in the CodeBuild User Guide.

The status of a build triggered by a webhook is always reported to your source provider.

startBuild_insecureSslOverride :: Lens' StartBuild (Maybe Bool)

Enable this flag to override the insecure SSL setting that is specified in the build project. The insecure SSL setting determines whether to ignore SSL warnings while connecting to the project source code. This override applies only if the build's source is GitHub Enterprise.

startBuild_imageOverride :: Lens' StartBuild (Maybe Text)

The name of an image for this build that overrides the one specified in the build project.

startBuild_secondaryArtifactsOverride :: Lens' StartBuild (Maybe [ProjectArtifacts])

An array of ProjectArtifacts objects.

startBuild_artifactsOverride :: Lens' StartBuild (Maybe ProjectArtifacts)

Build output artifact settings that override, for this build only, the latest ones already defined in the build project.

startBuild_sourceTypeOverride :: Lens' StartBuild (Maybe SourceType)

A source input type, for this build, that overrides the source input defined in the build project.

startBuild_projectName :: Lens' StartBuild Text

The name of the CodeBuild build project to start running a build.

Destructuring the Response

data StartBuildResponse

See: newStartBuildResponse smart constructor.

Constructors

StartBuildResponse'

Fields

build :: Maybe Build

Information about the build to be run.

httpStatus :: Int

The response's http status code.

newStartBuildResponse

Arguments

:: Int
$sel:httpStatus:StartBuildResponse'

-> StartBuildResponse

Create a value of StartBuildResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:build:StartBuildResponse', startBuildResponse_build - Information about the build to be run.

$sel:httpStatus:StartBuildResponse', startBuildResponse_httpStatus - The response's http status code.

Response Lenses

startBuildResponse_build :: Lens' StartBuildResponse (Maybe Build)

Information about the build to be run.

startBuildResponse_httpStatus :: Lens' StartBuildResponse Int

The response's http status code.