Amazonka.CodeBuild.Types.BuildArtifacts

Description

Documentation

data BuildArtifacts

Information about build output artifacts.

See: newBuildArtifacts smart constructor.

Constructors

BuildArtifacts'

Fields

location :: Maybe Text

Information about the location of the build artifacts.

md5sum :: Maybe Text

The MD5 hash of the build artifact.

You can use this hash along with a checksum tool to confirm file integrity and authenticity.

This value is available only if the build project's packaging value is set to ZIP.

encryptionDisabled :: Maybe Bool

Information that tells you if encryption for build artifacts is disabled.

overrideArtifactName :: Maybe Bool

If this flag is set, a name specified in the buildspec file overrides the artifact name. The name specified in a buildspec file is calculated at build time and uses the Shell Command Language. For example, you can append a date and time to your artifact name so that it is always unique.

artifactIdentifier :: Maybe Text

An identifier for this artifact definition.

sha256sum :: Maybe Text

The SHA-256 hash of the build artifact.

You can use this hash along with a checksum tool to confirm file integrity and authenticity.

This value is available only if the build project's packaging value is set to ZIP.

bucketOwnerAccess :: Maybe BucketOwnerAccess

newBuildArtifacts :: BuildArtifacts

Create a value of BuildArtifacts with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:location:BuildArtifacts', buildArtifacts_location - Information about the location of the build artifacts.

$sel:md5sum:BuildArtifacts', buildArtifacts_md5sum - The MD5 hash of the build artifact.

You can use this hash along with a checksum tool to confirm file integrity and authenticity.

This value is available only if the build project's packaging value is set to ZIP.

$sel:encryptionDisabled:BuildArtifacts', buildArtifacts_encryptionDisabled - Information that tells you if encryption for build artifacts is disabled.

$sel:overrideArtifactName:BuildArtifacts', buildArtifacts_overrideArtifactName - If this flag is set, a name specified in the buildspec file overrides the artifact name. The name specified in a buildspec file is calculated at build time and uses the Shell Command Language. For example, you can append a date and time to your artifact name so that it is always unique.

$sel:artifactIdentifier:BuildArtifacts', buildArtifacts_artifactIdentifier - An identifier for this artifact definition.

$sel:sha256sum:BuildArtifacts', buildArtifacts_sha256sum - The SHA-256 hash of the build artifact.

You can use this hash along with a checksum tool to confirm file integrity and authenticity.

This value is available only if the build project's packaging value is set to ZIP.

$sel:bucketOwnerAccess:BuildArtifacts', buildArtifacts_bucketOwnerAccess - Undocumented member.

buildArtifacts_location :: Lens' BuildArtifacts (Maybe Text)

Information about the location of the build artifacts.

buildArtifacts_md5sum :: Lens' BuildArtifacts (Maybe Text)

The MD5 hash of the build artifact.

You can use this hash along with a checksum tool to confirm file integrity and authenticity.

This value is available only if the build project's packaging value is set to ZIP.

buildArtifacts_encryptionDisabled :: Lens' BuildArtifacts (Maybe Bool)

Information that tells you if encryption for build artifacts is disabled.

buildArtifacts_overrideArtifactName :: Lens' BuildArtifacts (Maybe Bool)

If this flag is set, a name specified in the buildspec file overrides the artifact name. The name specified in a buildspec file is calculated at build time and uses the Shell Command Language. For example, you can append a date and time to your artifact name so that it is always unique.

buildArtifacts_artifactIdentifier :: Lens' BuildArtifacts (Maybe Text)

An identifier for this artifact definition.

buildArtifacts_sha256sum :: Lens' BuildArtifacts (Maybe Text)

The SHA-256 hash of the build artifact.

You can use this hash along with a checksum tool to confirm file integrity and authenticity.

This value is available only if the build project's packaging value is set to ZIP.

buildArtifacts_bucketOwnerAccess :: Lens' BuildArtifacts (Maybe BucketOwnerAccess)

Undocumented member.