Amazonka.CodeBuild.Types.ProjectSourceVersion

Description

Documentation

data ProjectSourceVersion

A source identifier and its corresponding version.

See: newProjectSourceVersion smart constructor.

Constructors

ProjectSourceVersion'

Fields

sourceIdentifier :: Text

An identifier for a source in the build project. The identifier can only contain alphanumeric characters and underscores, and must be less than 128 characters in length.

sourceVersion :: Text

The source version for the corresponding source identifier. If specified, must be one of:

For CodeCommit: the commit ID, branch, or Git tag to use.

For GitHub: the commit ID, pull request ID, branch name, or tag name that corresponds to the version of the source code you want to build. If a pull request ID is specified, it must use the format pr/pull-request-ID (for example, pr/25). If a branch name is specified, the branch's HEAD commit ID is used. If not specified, the default branch's HEAD commit ID is used.

For Bitbucket: the commit ID, branch name, or tag name that corresponds to the version of the source code you want to build. If a branch name is specified, the branch's HEAD commit ID is used. If not specified, the default branch's HEAD commit ID is used.

For Amazon S3: the version ID of the object that represents the build input ZIP file to use.

For more information, see Source Version Sample with CodeBuild in the CodeBuild User Guide.

newProjectSourceVersion

Arguments

:: Text
$sel:sourceIdentifier:ProjectSourceVersion'

-> Text
$sel:sourceVersion:ProjectSourceVersion'

-> ProjectSourceVersion

Create a value of ProjectSourceVersion with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:sourceIdentifier:ProjectSourceVersion', projectSourceVersion_sourceIdentifier - An identifier for a source in the build project. The identifier can only contain alphanumeric characters and underscores, and must be less than 128 characters in length.

$sel:sourceVersion:ProjectSourceVersion', projectSourceVersion_sourceVersion - The source version for the corresponding source identifier. If specified, must be one of:

For CodeCommit: the commit ID, branch, or Git tag to use.

For GitHub: the commit ID, pull request ID, branch name, or tag name that corresponds to the version of the source code you want to build. If a pull request ID is specified, it must use the format pr/pull-request-ID (for example, pr/25). If a branch name is specified, the branch's HEAD commit ID is used. If not specified, the default branch's HEAD commit ID is used.

For Bitbucket: the commit ID, branch name, or tag name that corresponds to the version of the source code you want to build. If a branch name is specified, the branch's HEAD commit ID is used. If not specified, the default branch's HEAD commit ID is used.

For Amazon S3: the version ID of the object that represents the build input ZIP file to use.

For more information, see Source Version Sample with CodeBuild in the CodeBuild User Guide.

projectSourceVersion_sourceIdentifier :: Lens' ProjectSourceVersion Text

An identifier for a source in the build project. The identifier can only contain alphanumeric characters and underscores, and must be less than 128 characters in length.

projectSourceVersion_sourceVersion :: Lens' ProjectSourceVersion Text

The source version for the corresponding source identifier. If specified, must be one of:

For CodeCommit: the commit ID, branch, or Git tag to use.

For GitHub: the commit ID, pull request ID, branch name, or tag name that corresponds to the version of the source code you want to build. If a pull request ID is specified, it must use the format pr/pull-request-ID (for example, pr/25). If a branch name is specified, the branch's HEAD commit ID is used. If not specified, the default branch's HEAD commit ID is used.

For Bitbucket: the commit ID, branch name, or tag name that corresponds to the version of the source code you want to build. If a branch name is specified, the branch's HEAD commit ID is used. If not specified, the default branch's HEAD commit ID is used.

For Amazon S3: the version ID of the object that represents the build input ZIP file to use.

For more information, see Source Version Sample with CodeBuild in the CodeBuild User Guide.