Amazonka.CodeBuild.Types.ReportGroup

Description

Documentation

data ReportGroup

A series of reports. Each report contains information about the results from running a series of test cases. You specify the test cases for a report group in the buildspec for a build project using one or more paths to the test case files.

See: newReportGroup smart constructor.

Constructors

ReportGroup'

Fields

status :: Maybe ReportGroupStatusType

The status of the report group. This property is read-only.

This can be one of the following values:

ACTIVE

The report group is active.

DELETING

The report group is in the process of being deleted.

arn :: Maybe Text

The ARN of the ReportGroup.

created :: Maybe POSIX

The date and time this ReportGroup was created.

name :: Maybe Text

The name of the ReportGroup.

type' :: Maybe ReportType

The type of the ReportGroup. This can be one of the following values:

CODE_COVERAGE

The report group contains code coverage reports.

TEST

The report group contains test reports.

lastModified :: Maybe POSIX

The date and time this ReportGroup was last modified.

exportConfig :: Maybe ReportExportConfig

Information about the destination where the raw data of this ReportGroup is exported.

tags :: Maybe [Tag]

A list of tag key and value pairs associated with this report group.

These tags are available for use by Amazon Web Services services that support CodeBuild report group tags.

newReportGroup :: ReportGroup

Create a value of ReportGroup with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:status:ReportGroup', reportGroup_status - The status of the report group. This property is read-only.

This can be one of the following values:

ACTIVE

The report group is active.

DELETING

The report group is in the process of being deleted.

$sel:arn:ReportGroup', reportGroup_arn - The ARN of the ReportGroup.

$sel:created:ReportGroup', reportGroup_created - The date and time this ReportGroup was created.

$sel:name:ReportGroup', reportGroup_name - The name of the ReportGroup.

$sel:type':ReportGroup', reportGroup_type - The type of the ReportGroup. This can be one of the following values:

CODE_COVERAGE

The report group contains code coverage reports.

TEST

The report group contains test reports.

$sel:lastModified:ReportGroup', reportGroup_lastModified - The date and time this ReportGroup was last modified.

$sel:exportConfig:ReportGroup', reportGroup_exportConfig - Information about the destination where the raw data of this ReportGroup is exported.

$sel:tags:ReportGroup', reportGroup_tags - A list of tag key and value pairs associated with this report group.

These tags are available for use by Amazon Web Services services that support CodeBuild report group tags.

reportGroup_status :: Lens' ReportGroup (Maybe ReportGroupStatusType)

The status of the report group. This property is read-only.

This can be one of the following values:

ACTIVE

The report group is active.

DELETING

The report group is in the process of being deleted.

reportGroup_arn :: Lens' ReportGroup (Maybe Text)

The ARN of the ReportGroup.

reportGroup_created :: Lens' ReportGroup (Maybe UTCTime)

The date and time this ReportGroup was created.

reportGroup_name :: Lens' ReportGroup (Maybe Text)

The name of the ReportGroup.

reportGroup_type :: Lens' ReportGroup (Maybe ReportType)

The type of the ReportGroup. This can be one of the following values:

CODE_COVERAGE

The report group contains code coverage reports.

TEST

The report group contains test reports.

reportGroup_lastModified :: Lens' ReportGroup (Maybe UTCTime)

The date and time this ReportGroup was last modified.

reportGroup_exportConfig :: Lens' ReportGroup (Maybe ReportExportConfig)

Information about the destination where the raw data of this ReportGroup is exported.

reportGroup_tags :: Lens' ReportGroup (Maybe [Tag])

A list of tag key and value pairs associated with this report group.

These tags are available for use by Amazon Web Services services that support CodeBuild report group tags.