Amazonka.CodeBuild.Types.ResolvedArtifact

Description

Documentation

data ResolvedArtifact

Represents a resolved build artifact. A resolved artifact is an artifact that is built and deployed to the destination, such as Amazon S3.

See: newResolvedArtifact smart constructor.

Constructors

ResolvedArtifact'

Fields

location :: Maybe Text

The location of the artifact.

identifier :: Maybe Text

The identifier of the artifact.

type' :: Maybe ArtifactsType

Specifies the type of artifact.

newResolvedArtifact :: ResolvedArtifact

Create a value of ResolvedArtifact with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:location:ResolvedArtifact', resolvedArtifact_location - The location of the artifact.

$sel:identifier:ResolvedArtifact', resolvedArtifact_identifier - The identifier of the artifact.

$sel:type':ResolvedArtifact', resolvedArtifact_type - Specifies the type of artifact.

resolvedArtifact_location :: Lens' ResolvedArtifact (Maybe Text)

The location of the artifact.

resolvedArtifact_identifier :: Lens' ResolvedArtifact (Maybe Text)

The identifier of the artifact.

resolvedArtifact_type :: Lens' ResolvedArtifact (Maybe ArtifactsType)

Specifies the type of artifact.