Amazonka.CodeBuild.UpdateProject

Description

Changes the settings of a build project.

Creating a Request

data UpdateProject

See: newUpdateProject smart constructor.

Constructors

UpdateProject'

Fields

secondaryArtifacts :: Maybe [ProjectArtifacts]

An array of ProjectArtifact objects.

artifacts :: Maybe ProjectArtifacts

Information to be changed about the build output artifacts for the build project.

environment :: Maybe ProjectEnvironment

Information to be changed about the build environment for the build project.

concurrentBuildLimit :: Maybe Int

The maximum number of concurrent builds that are allowed for this project.

New builds are only started if the current number of builds is less than or equal to this limit. If the current build count meets this limit, new builds are throttled and are not run.

To remove this limit, set this value to -1.

badgeEnabled :: Maybe Bool

Set this to true to generate a publicly accessible URL for your project's build badge.

secondarySourceVersions :: Maybe [ProjectSourceVersion]

An array of ProjectSourceVersion objects. If secondarySourceVersions is specified at the build level, then they take over these secondarySourceVersions (at the project level).

queuedTimeoutInMinutes :: Maybe Natural

The number of minutes a build is allowed to be queued before it times out.

cache :: Maybe ProjectCache

Stores recently used information so that it can be quickly accessed at a later time.

secondarySources :: Maybe [ProjectSource]

An array of ProjectSource objects.

sourceVersion :: Maybe Text

A version of the build input to be built for this project. If not specified, the latest version is used. If specified, it must be one of:

For CodeCommit: the commit ID, branch, or Git tag to use.

For GitHub: the commit ID, pull request ID, branch name, or tag name that corresponds to the version of the source code you want to build. If a pull request ID is specified, it must use the format pr/pull-request-ID (for example pr/25). If a branch name is specified, the branch's HEAD commit ID is used. If not specified, the default branch's HEAD commit ID is used.

For Bitbucket: the commit ID, branch name, or tag name that corresponds to the version of the source code you want to build. If a branch name is specified, the branch's HEAD commit ID is used. If not specified, the default branch's HEAD commit ID is used.

For Amazon S3: the version ID of the object that represents the build input ZIP file to use.

If sourceVersion is specified at the build level, then that version takes precedence over this sourceVersion (at the project level).

For more information, see Source Version Sample with CodeBuild in the CodeBuild User Guide.

vpcConfig :: Maybe VpcConfig

VpcConfig enables CodeBuild to access resources in an Amazon VPC.

source :: Maybe ProjectSource

Information to be changed about the build input source code for the build project.

logsConfig :: Maybe LogsConfig

Information about logs for the build project. A project can create logs in CloudWatch Logs, logs in an S3 bucket, or both.

fileSystemLocations :: Maybe [ProjectFileSystemLocation]

An array of ProjectFileSystemLocation objects for a CodeBuild build project. A ProjectFileSystemLocation object specifies the identifier, location, mountOptions, mountPoint, and type of a file system created using Amazon Elastic File System.

buildBatchConfig :: Maybe ProjectBuildBatchConfig

encryptionKey :: Maybe Text

The Key Management Service customer master key (CMK) to be used for encrypting the build output artifacts.

You can use a cross-account KMS key to encrypt the build output artifacts if your service role has permission to that key.

You can specify either the Amazon Resource Name (ARN) of the CMK or, if available, the CMK's alias (using the format alias/<alias-name>).

description :: Maybe Text

A new or replacement description of the build project.

serviceRole :: Maybe Text

The replacement ARN of the IAM role that enables CodeBuild to interact with dependent Amazon Web Services services on behalf of the Amazon Web Services account.

tags :: Maybe [Tag]

An updated list of tag key and value pairs associated with this build project.

These tags are available for use by Amazon Web Services services that support CodeBuild build project tags.

timeoutInMinutes :: Maybe Natural

The replacement value in minutes, from 5 to 480 (8 hours), for CodeBuild to wait before timing out any related build that did not get marked as completed.

name :: Text

The name of the build project.

You cannot change a build project's name.

newUpdateProject

Arguments

:: Text
$sel:name:UpdateProject'

-> UpdateProject

Create a value of UpdateProject with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:secondaryArtifacts:UpdateProject', updateProject_secondaryArtifacts - An array of ProjectArtifact objects.

$sel:artifacts:UpdateProject', updateProject_artifacts - Information to be changed about the build output artifacts for the build project.

$sel:environment:UpdateProject', updateProject_environment - Information to be changed about the build environment for the build project.

$sel:concurrentBuildLimit:UpdateProject', updateProject_concurrentBuildLimit - The maximum number of concurrent builds that are allowed for this project.

New builds are only started if the current number of builds is less than or equal to this limit. If the current build count meets this limit, new builds are throttled and are not run.

To remove this limit, set this value to -1.

$sel:badgeEnabled:UpdateProject', updateProject_badgeEnabled - Set this to true to generate a publicly accessible URL for your project's build badge.

$sel:secondarySourceVersions:UpdateProject', updateProject_secondarySourceVersions - An array of ProjectSourceVersion objects. If secondarySourceVersions is specified at the build level, then they take over these secondarySourceVersions (at the project level).

$sel:queuedTimeoutInMinutes:UpdateProject', updateProject_queuedTimeoutInMinutes - The number of minutes a build is allowed to be queued before it times out.

$sel:cache:UpdateProject', updateProject_cache - Stores recently used information so that it can be quickly accessed at a later time.

$sel:secondarySources:UpdateProject', updateProject_secondarySources - An array of ProjectSource objects.

$sel:sourceVersion:UpdateProject', updateProject_sourceVersion - A version of the build input to be built for this project. If not specified, the latest version is used. If specified, it must be one of:

For CodeCommit: the commit ID, branch, or Git tag to use.

For GitHub: the commit ID, pull request ID, branch name, or tag name that corresponds to the version of the source code you want to build. If a pull request ID is specified, it must use the format pr/pull-request-ID (for example pr/25). If a branch name is specified, the branch's HEAD commit ID is used. If not specified, the default branch's HEAD commit ID is used.

For Bitbucket: the commit ID, branch name, or tag name that corresponds to the version of the source code you want to build. If a branch name is specified, the branch's HEAD commit ID is used. If not specified, the default branch's HEAD commit ID is used.

For Amazon S3: the version ID of the object that represents the build input ZIP file to use.

If sourceVersion is specified at the build level, then that version takes precedence over this sourceVersion (at the project level).

For more information, see Source Version Sample with CodeBuild in the CodeBuild User Guide.

$sel:vpcConfig:UpdateProject', updateProject_vpcConfig - VpcConfig enables CodeBuild to access resources in an Amazon VPC.

$sel:source:UpdateProject', updateProject_source - Information to be changed about the build input source code for the build project.

$sel:logsConfig:UpdateProject', updateProject_logsConfig - Information about logs for the build project. A project can create logs in CloudWatch Logs, logs in an S3 bucket, or both.

$sel:fileSystemLocations:UpdateProject', updateProject_fileSystemLocations - An array of ProjectFileSystemLocation objects for a CodeBuild build project. A ProjectFileSystemLocation object specifies the identifier, location, mountOptions, mountPoint, and type of a file system created using Amazon Elastic File System.

$sel:buildBatchConfig:UpdateProject', updateProject_buildBatchConfig - Undocumented member.

$sel:encryptionKey:UpdateProject', updateProject_encryptionKey - The Key Management Service customer master key (CMK) to be used for encrypting the build output artifacts.

You can use a cross-account KMS key to encrypt the build output artifacts if your service role has permission to that key.

You can specify either the Amazon Resource Name (ARN) of the CMK or, if available, the CMK's alias (using the format alias/<alias-name>).

$sel:description:UpdateProject', updateProject_description - A new or replacement description of the build project.

$sel:serviceRole:UpdateProject', updateProject_serviceRole - The replacement ARN of the IAM role that enables CodeBuild to interact with dependent Amazon Web Services services on behalf of the Amazon Web Services account.

$sel:tags:UpdateProject', updateProject_tags - An updated list of tag key and value pairs associated with this build project.

These tags are available for use by Amazon Web Services services that support CodeBuild build project tags.

$sel:timeoutInMinutes:UpdateProject', updateProject_timeoutInMinutes - The replacement value in minutes, from 5 to 480 (8 hours), for CodeBuild to wait before timing out any related build that did not get marked as completed.

$sel:name:UpdateProject', updateProject_name - The name of the build project.

You cannot change a build project's name.

Request Lenses

updateProject_secondaryArtifacts :: Lens' UpdateProject (Maybe [ProjectArtifacts])

An array of ProjectArtifact objects.

updateProject_artifacts :: Lens' UpdateProject (Maybe ProjectArtifacts)

Information to be changed about the build output artifacts for the build project.

updateProject_environment :: Lens' UpdateProject (Maybe ProjectEnvironment)

Information to be changed about the build environment for the build project.

updateProject_concurrentBuildLimit :: Lens' UpdateProject (Maybe Int)

The maximum number of concurrent builds that are allowed for this project.

New builds are only started if the current number of builds is less than or equal to this limit. If the current build count meets this limit, new builds are throttled and are not run.

To remove this limit, set this value to -1.

updateProject_badgeEnabled :: Lens' UpdateProject (Maybe Bool)

Set this to true to generate a publicly accessible URL for your project's build badge.

updateProject_secondarySourceVersions :: Lens' UpdateProject (Maybe [ProjectSourceVersion])

An array of ProjectSourceVersion objects. If secondarySourceVersions is specified at the build level, then they take over these secondarySourceVersions (at the project level).

updateProject_queuedTimeoutInMinutes :: Lens' UpdateProject (Maybe Natural)

The number of minutes a build is allowed to be queued before it times out.

updateProject_cache :: Lens' UpdateProject (Maybe ProjectCache)

Stores recently used information so that it can be quickly accessed at a later time.

updateProject_secondarySources :: Lens' UpdateProject (Maybe [ProjectSource])

An array of ProjectSource objects.

updateProject_sourceVersion :: Lens' UpdateProject (Maybe Text)

A version of the build input to be built for this project. If not specified, the latest version is used. If specified, it must be one of:

For CodeCommit: the commit ID, branch, or Git tag to use.

For GitHub: the commit ID, pull request ID, branch name, or tag name that corresponds to the version of the source code you want to build. If a pull request ID is specified, it must use the format pr/pull-request-ID (for example pr/25). If a branch name is specified, the branch's HEAD commit ID is used. If not specified, the default branch's HEAD commit ID is used.

For Bitbucket: the commit ID, branch name, or tag name that corresponds to the version of the source code you want to build. If a branch name is specified, the branch's HEAD commit ID is used. If not specified, the default branch's HEAD commit ID is used.

For Amazon S3: the version ID of the object that represents the build input ZIP file to use.

If sourceVersion is specified at the build level, then that version takes precedence over this sourceVersion (at the project level).

For more information, see Source Version Sample with CodeBuild in the CodeBuild User Guide.

updateProject_vpcConfig :: Lens' UpdateProject (Maybe VpcConfig)

VpcConfig enables CodeBuild to access resources in an Amazon VPC.

updateProject_source :: Lens' UpdateProject (Maybe ProjectSource)

Information to be changed about the build input source code for the build project.

updateProject_logsConfig :: Lens' UpdateProject (Maybe LogsConfig)

Information about logs for the build project. A project can create logs in CloudWatch Logs, logs in an S3 bucket, or both.

updateProject_fileSystemLocations :: Lens' UpdateProject (Maybe [ProjectFileSystemLocation])

An array of ProjectFileSystemLocation objects for a CodeBuild build project. A ProjectFileSystemLocation object specifies the identifier, location, mountOptions, mountPoint, and type of a file system created using Amazon Elastic File System.

updateProject_buildBatchConfig :: Lens' UpdateProject (Maybe ProjectBuildBatchConfig)

Undocumented member.

updateProject_encryptionKey :: Lens' UpdateProject (Maybe Text)

The Key Management Service customer master key (CMK) to be used for encrypting the build output artifacts.

You can use a cross-account KMS key to encrypt the build output artifacts if your service role has permission to that key.

You can specify either the Amazon Resource Name (ARN) of the CMK or, if available, the CMK's alias (using the format alias/<alias-name>).

updateProject_description :: Lens' UpdateProject (Maybe Text)

A new or replacement description of the build project.

updateProject_serviceRole :: Lens' UpdateProject (Maybe Text)

The replacement ARN of the IAM role that enables CodeBuild to interact with dependent Amazon Web Services services on behalf of the Amazon Web Services account.

updateProject_tags :: Lens' UpdateProject (Maybe [Tag])

An updated list of tag key and value pairs associated with this build project.

These tags are available for use by Amazon Web Services services that support CodeBuild build project tags.

updateProject_timeoutInMinutes :: Lens' UpdateProject (Maybe Natural)

The replacement value in minutes, from 5 to 480 (8 hours), for CodeBuild to wait before timing out any related build that did not get marked as completed.

updateProject_name :: Lens' UpdateProject Text

The name of the build project.

You cannot change a build project's name.

Destructuring the Response

data UpdateProjectResponse

See: newUpdateProjectResponse smart constructor.

Constructors

UpdateProjectResponse'

Fields

project :: Maybe Project

Information about the build project that was changed.

httpStatus :: Int

The response's http status code.

newUpdateProjectResponse

Arguments

:: Int
$sel:httpStatus:UpdateProjectResponse'

-> UpdateProjectResponse

Create a value of UpdateProjectResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:project:UpdateProjectResponse', updateProjectResponse_project - Information about the build project that was changed.

$sel:httpStatus:UpdateProjectResponse', updateProjectResponse_httpStatus - The response's http status code.

Response Lenses

updateProjectResponse_project :: Lens' UpdateProjectResponse (Maybe Project)

Information about the build project that was changed.

updateProjectResponse_httpStatus :: Lens' UpdateProjectResponse Int

The response's http status code.