Amazonka.CodeCommit.BatchDescribeMergeConflicts

Description

Returns information about one or more merge conflicts in the attempted merge of two commit specifiers using the squash or three-way merge strategy.

Creating a Request

data BatchDescribeMergeConflicts

See: newBatchDescribeMergeConflicts smart constructor.

Constructors

BatchDescribeMergeConflicts'

Fields

filePaths :: Maybe [Text]

The path of the target files used to describe the conflicts. If not specified, the default is all conflict files.

conflictDetailLevel :: Maybe ConflictDetailLevelTypeEnum

The level of conflict detail to use. If unspecified, the default FILE_LEVEL is used, which returns a not-mergeable result if the same file has differences in both branches. If LINE_LEVEL is specified, a conflict is considered not mergeable if the same file in both branches has differences on the same line.

nextToken :: Maybe Text

An enumeration token that, when provided in a request, returns the next batch of the results.

maxConflictFiles :: Maybe Int

The maximum number of files to include in the output.

maxMergeHunks :: Maybe Int

The maximum number of merge hunks to include in the output.

conflictResolutionStrategy :: Maybe ConflictResolutionStrategyTypeEnum

Specifies which branch to use when resolving conflicts, or whether to attempt automatically merging two versions of a file. The default is NONE, which requires any conflicts to be resolved manually before the merge operation is successful.

repositoryName :: Text

The name of the repository that contains the merge conflicts you want to review.

destinationCommitSpecifier :: Text

The branch, tag, HEAD, or other fully qualified reference used to identify a commit (for example, a branch name or a full commit ID).

sourceCommitSpecifier :: Text

The branch, tag, HEAD, or other fully qualified reference used to identify a commit (for example, a branch name or a full commit ID).

mergeOption :: MergeOptionTypeEnum

The merge option or strategy you want to use to merge the code.

newBatchDescribeMergeConflicts

Arguments

:: Text
$sel:repositoryName:BatchDescribeMergeConflicts'

-> Text
$sel:destinationCommitSpecifier:BatchDescribeMergeConflicts'

-> Text
$sel:sourceCommitSpecifier:BatchDescribeMergeConflicts'

-> MergeOptionTypeEnum
$sel:mergeOption:BatchDescribeMergeConflicts'

-> BatchDescribeMergeConflicts

Create a value of BatchDescribeMergeConflicts with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:filePaths:BatchDescribeMergeConflicts', batchDescribeMergeConflicts_filePaths - The path of the target files used to describe the conflicts. If not specified, the default is all conflict files.

$sel:conflictDetailLevel:BatchDescribeMergeConflicts', batchDescribeMergeConflicts_conflictDetailLevel - The level of conflict detail to use. If unspecified, the default FILE_LEVEL is used, which returns a not-mergeable result if the same file has differences in both branches. If LINE_LEVEL is specified, a conflict is considered not mergeable if the same file in both branches has differences on the same line.

$sel:nextToken:BatchDescribeMergeConflicts', batchDescribeMergeConflicts_nextToken - An enumeration token that, when provided in a request, returns the next batch of the results.

$sel:maxConflictFiles:BatchDescribeMergeConflicts', batchDescribeMergeConflicts_maxConflictFiles - The maximum number of files to include in the output.

$sel:maxMergeHunks:BatchDescribeMergeConflicts', batchDescribeMergeConflicts_maxMergeHunks - The maximum number of merge hunks to include in the output.

$sel:conflictResolutionStrategy:BatchDescribeMergeConflicts', batchDescribeMergeConflicts_conflictResolutionStrategy - Specifies which branch to use when resolving conflicts, or whether to attempt automatically merging two versions of a file. The default is NONE, which requires any conflicts to be resolved manually before the merge operation is successful.

$sel:repositoryName:BatchDescribeMergeConflicts', batchDescribeMergeConflicts_repositoryName - The name of the repository that contains the merge conflicts you want to review.

$sel:destinationCommitSpecifier:BatchDescribeMergeConflicts', batchDescribeMergeConflicts_destinationCommitSpecifier - The branch, tag, HEAD, or other fully qualified reference used to identify a commit (for example, a branch name or a full commit ID).

$sel:sourceCommitSpecifier:BatchDescribeMergeConflicts', batchDescribeMergeConflicts_sourceCommitSpecifier - The branch, tag, HEAD, or other fully qualified reference used to identify a commit (for example, a branch name or a full commit ID).

$sel:mergeOption:BatchDescribeMergeConflicts', batchDescribeMergeConflicts_mergeOption - The merge option or strategy you want to use to merge the code.

Request Lenses

batchDescribeMergeConflicts_filePaths :: Lens' BatchDescribeMergeConflicts (Maybe [Text])

The path of the target files used to describe the conflicts. If not specified, the default is all conflict files.

batchDescribeMergeConflicts_conflictDetailLevel :: Lens' BatchDescribeMergeConflicts (Maybe ConflictDetailLevelTypeEnum)

The level of conflict detail to use. If unspecified, the default FILE_LEVEL is used, which returns a not-mergeable result if the same file has differences in both branches. If LINE_LEVEL is specified, a conflict is considered not mergeable if the same file in both branches has differences on the same line.

batchDescribeMergeConflicts_nextToken :: Lens' BatchDescribeMergeConflicts (Maybe Text)

An enumeration token that, when provided in a request, returns the next batch of the results.

batchDescribeMergeConflicts_maxConflictFiles :: Lens' BatchDescribeMergeConflicts (Maybe Int)

The maximum number of files to include in the output.

batchDescribeMergeConflicts_maxMergeHunks :: Lens' BatchDescribeMergeConflicts (Maybe Int)

The maximum number of merge hunks to include in the output.

batchDescribeMergeConflicts_conflictResolutionStrategy :: Lens' BatchDescribeMergeConflicts (Maybe ConflictResolutionStrategyTypeEnum)

Specifies which branch to use when resolving conflicts, or whether to attempt automatically merging two versions of a file. The default is NONE, which requires any conflicts to be resolved manually before the merge operation is successful.

batchDescribeMergeConflicts_repositoryName :: Lens' BatchDescribeMergeConflicts Text

The name of the repository that contains the merge conflicts you want to review.

batchDescribeMergeConflicts_destinationCommitSpecifier :: Lens' BatchDescribeMergeConflicts Text

The branch, tag, HEAD, or other fully qualified reference used to identify a commit (for example, a branch name or a full commit ID).

batchDescribeMergeConflicts_sourceCommitSpecifier :: Lens' BatchDescribeMergeConflicts Text

The branch, tag, HEAD, or other fully qualified reference used to identify a commit (for example, a branch name or a full commit ID).

batchDescribeMergeConflicts_mergeOption :: Lens' BatchDescribeMergeConflicts MergeOptionTypeEnum

The merge option or strategy you want to use to merge the code.

Destructuring the Response

data BatchDescribeMergeConflictsResponse

See: newBatchDescribeMergeConflictsResponse smart constructor.

Constructors

BatchDescribeMergeConflictsResponse'

Fields

baseCommitId :: Maybe Text

The commit ID of the merge base.

nextToken :: Maybe Text

An enumeration token that can be used in a request to return the next batch of the results.

errors :: Maybe [BatchDescribeMergeConflictsError]

A list of any errors returned while describing the merge conflicts for each file.

httpStatus :: Int

The response's http status code.

conflicts :: [Conflict]

A list of conflicts for each file, including the conflict metadata and the hunks of the differences between the files.

destinationCommitId :: Text

The commit ID of the destination commit specifier that was used in the merge evaluation.

sourceCommitId :: Text

The commit ID of the source commit specifier that was used in the merge evaluation.

newBatchDescribeMergeConflictsResponse

Arguments

:: Int
$sel:httpStatus:BatchDescribeMergeConflictsResponse'

-> Text
$sel:destinationCommitId:BatchDescribeMergeConflictsResponse'

-> Text
$sel:sourceCommitId:BatchDescribeMergeConflictsResponse'

-> BatchDescribeMergeConflictsResponse

Create a value of BatchDescribeMergeConflictsResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:baseCommitId:BatchDescribeMergeConflictsResponse', batchDescribeMergeConflictsResponse_baseCommitId - The commit ID of the merge base.

$sel:nextToken:BatchDescribeMergeConflicts', batchDescribeMergeConflictsResponse_nextToken - An enumeration token that can be used in a request to return the next batch of the results.

$sel:errors:BatchDescribeMergeConflictsResponse', batchDescribeMergeConflictsResponse_errors - A list of any errors returned while describing the merge conflicts for each file.

$sel:httpStatus:BatchDescribeMergeConflictsResponse', batchDescribeMergeConflictsResponse_httpStatus - The response's http status code.

$sel:conflicts:BatchDescribeMergeConflictsResponse', batchDescribeMergeConflictsResponse_conflicts - A list of conflicts for each file, including the conflict metadata and the hunks of the differences between the files.

$sel:destinationCommitId:BatchDescribeMergeConflictsResponse', batchDescribeMergeConflictsResponse_destinationCommitId - The commit ID of the destination commit specifier that was used in the merge evaluation.

$sel:sourceCommitId:BatchDescribeMergeConflictsResponse', batchDescribeMergeConflictsResponse_sourceCommitId - The commit ID of the source commit specifier that was used in the merge evaluation.

Response Lenses

batchDescribeMergeConflictsResponse_baseCommitId :: Lens' BatchDescribeMergeConflictsResponse (Maybe Text)

The commit ID of the merge base.

batchDescribeMergeConflictsResponse_nextToken :: Lens' BatchDescribeMergeConflictsResponse (Maybe Text)

An enumeration token that can be used in a request to return the next batch of the results.

batchDescribeMergeConflictsResponse_errors :: Lens' BatchDescribeMergeConflictsResponse (Maybe [BatchDescribeMergeConflictsError])

A list of any errors returned while describing the merge conflicts for each file.

batchDescribeMergeConflictsResponse_httpStatus :: Lens' BatchDescribeMergeConflictsResponse Int

The response's http status code.

batchDescribeMergeConflictsResponse_conflicts :: Lens' BatchDescribeMergeConflictsResponse [Conflict]

A list of conflicts for each file, including the conflict metadata and the hunks of the differences between the files.

batchDescribeMergeConflictsResponse_destinationCommitId :: Lens' BatchDescribeMergeConflictsResponse Text

The commit ID of the destination commit specifier that was used in the merge evaluation.

batchDescribeMergeConflictsResponse_sourceCommitId :: Lens' BatchDescribeMergeConflictsResponse Text

The commit ID of the source commit specifier that was used in the merge evaluation.