Amazonka.CodeCommit.BatchGetRepositories

Description

Returns information about one or more repositories.

The description field for a repository accepts all HTML characters and all valid Unicode characters. Applications that do not HTML-encode the description and display it in a webpage can expose users to potentially malicious code. Make sure that you HTML-encode the description field in any application that uses this API to display the repository description on a webpage.

Creating a Request

data BatchGetRepositories

Represents the input of a batch get repositories operation.

See: newBatchGetRepositories smart constructor.

Constructors

BatchGetRepositories'

Fields

repositoryNames :: [Text]

The names of the repositories to get information about.

The length constraint limit is for each string in the array. The array itself can be empty.

newBatchGetRepositories :: BatchGetRepositories

Create a value of BatchGetRepositories with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:repositoryNames:BatchGetRepositories', batchGetRepositories_repositoryNames - The names of the repositories to get information about.

The length constraint limit is for each string in the array. The array itself can be empty.

Request Lenses

batchGetRepositories_repositoryNames :: Lens' BatchGetRepositories [Text]

The names of the repositories to get information about.

The length constraint limit is for each string in the array. The array itself can be empty.

Destructuring the Response

data BatchGetRepositoriesResponse

Represents the output of a batch get repositories operation.

See: newBatchGetRepositoriesResponse smart constructor.

Constructors

BatchGetRepositoriesResponse'

Fields

repositories :: Maybe [RepositoryMetadata]

A list of repositories returned by the batch get repositories operation.

repositoriesNotFound :: Maybe [Text]

Returns a list of repository names for which information could not be found.

httpStatus :: Int

The response's http status code.

newBatchGetRepositoriesResponse

Arguments

:: Int
$sel:httpStatus:BatchGetRepositoriesResponse'

-> BatchGetRepositoriesResponse

Create a value of BatchGetRepositoriesResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:repositories:BatchGetRepositoriesResponse', batchGetRepositoriesResponse_repositories - A list of repositories returned by the batch get repositories operation.

$sel:repositoriesNotFound:BatchGetRepositoriesResponse', batchGetRepositoriesResponse_repositoriesNotFound - Returns a list of repository names for which information could not be found.

$sel:httpStatus:BatchGetRepositoriesResponse', batchGetRepositoriesResponse_httpStatus - The response's http status code.

Response Lenses

batchGetRepositoriesResponse_repositories :: Lens' BatchGetRepositoriesResponse (Maybe [RepositoryMetadata])

A list of repositories returned by the batch get repositories operation.

batchGetRepositoriesResponse_repositoriesNotFound :: Lens' BatchGetRepositoriesResponse (Maybe [Text])

Returns a list of repository names for which information could not be found.

batchGetRepositoriesResponse_httpStatus :: Lens' BatchGetRepositoriesResponse Int

The response's http status code.