Amazonka.CodeCommit.CreateUnreferencedMergeCommit

Description

Creates an unreferenced commit that represents the result of merging two branches using a specified merge strategy. This can help you determine the outcome of a potential merge. This API cannot be used with the fast-forward merge strategy because that strategy does not create a merge commit.

This unreferenced merge commit can only be accessed using the GetCommit API or through git commands such as git fetch. To retrieve this commit, you must specify its commit ID or otherwise reference it.

Creating a Request

data CreateUnreferencedMergeCommit

See: newCreateUnreferencedMergeCommit smart constructor.

Constructors

CreateUnreferencedMergeCommit'

Fields

email :: Maybe Text

The email address for the person who created the unreferenced commit.

authorName :: Maybe Text

The name of the author who created the unreferenced commit. This information is used as both the author and committer for the commit.

conflictDetailLevel :: Maybe ConflictDetailLevelTypeEnum

The level of conflict detail to use. If unspecified, the default FILE_LEVEL is used, which returns a not-mergeable result if the same file has differences in both branches. If LINE_LEVEL is specified, a conflict is considered not mergeable if the same file in both branches has differences on the same line.

commitMessage :: Maybe Text

The commit message for the unreferenced commit.

conflictResolution :: Maybe ConflictResolution

If AUTOMERGE is the conflict resolution strategy, a list of inputs to use when resolving conflicts during a merge.

conflictResolutionStrategy :: Maybe ConflictResolutionStrategyTypeEnum

Specifies which branch to use when resolving conflicts, or whether to attempt automatically merging two versions of a file. The default is NONE, which requires any conflicts to be resolved manually before the merge operation is successful.

keepEmptyFolders :: Maybe Bool

If the commit contains deletions, whether to keep a folder or folder structure if the changes leave the folders empty. If this is specified as true, a .gitkeep file is created for empty folders. The default is false.

repositoryName :: Text

The name of the repository where you want to create the unreferenced merge commit.

sourceCommitSpecifier :: Text

The branch, tag, HEAD, or other fully qualified reference used to identify a commit (for example, a branch name or a full commit ID).

destinationCommitSpecifier :: Text

The branch, tag, HEAD, or other fully qualified reference used to identify a commit (for example, a branch name or a full commit ID).

mergeOption :: MergeOptionTypeEnum

The merge option or strategy you want to use to merge the code.

newCreateUnreferencedMergeCommit

Arguments

:: Text
$sel:repositoryName:CreateUnreferencedMergeCommit'

-> Text
$sel:sourceCommitSpecifier:CreateUnreferencedMergeCommit'

-> Text
$sel:destinationCommitSpecifier:CreateUnreferencedMergeCommit'

-> MergeOptionTypeEnum
$sel:mergeOption:CreateUnreferencedMergeCommit'

-> CreateUnreferencedMergeCommit

Create a value of CreateUnreferencedMergeCommit with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:email:CreateUnreferencedMergeCommit', createUnreferencedMergeCommit_email - The email address for the person who created the unreferenced commit.

$sel:authorName:CreateUnreferencedMergeCommit', createUnreferencedMergeCommit_authorName - The name of the author who created the unreferenced commit. This information is used as both the author and committer for the commit.

$sel:conflictDetailLevel:CreateUnreferencedMergeCommit', createUnreferencedMergeCommit_conflictDetailLevel - The level of conflict detail to use. If unspecified, the default FILE_LEVEL is used, which returns a not-mergeable result if the same file has differences in both branches. If LINE_LEVEL is specified, a conflict is considered not mergeable if the same file in both branches has differences on the same line.

$sel:commitMessage:CreateUnreferencedMergeCommit', createUnreferencedMergeCommit_commitMessage - The commit message for the unreferenced commit.

$sel:conflictResolution:CreateUnreferencedMergeCommit', createUnreferencedMergeCommit_conflictResolution - If AUTOMERGE is the conflict resolution strategy, a list of inputs to use when resolving conflicts during a merge.

$sel:conflictResolutionStrategy:CreateUnreferencedMergeCommit', createUnreferencedMergeCommit_conflictResolutionStrategy - Specifies which branch to use when resolving conflicts, or whether to attempt automatically merging two versions of a file. The default is NONE, which requires any conflicts to be resolved manually before the merge operation is successful.

$sel:keepEmptyFolders:CreateUnreferencedMergeCommit', createUnreferencedMergeCommit_keepEmptyFolders - If the commit contains deletions, whether to keep a folder or folder structure if the changes leave the folders empty. If this is specified as true, a .gitkeep file is created for empty folders. The default is false.

$sel:repositoryName:CreateUnreferencedMergeCommit', createUnreferencedMergeCommit_repositoryName - The name of the repository where you want to create the unreferenced merge commit.

$sel:sourceCommitSpecifier:CreateUnreferencedMergeCommit', createUnreferencedMergeCommit_sourceCommitSpecifier - The branch, tag, HEAD, or other fully qualified reference used to identify a commit (for example, a branch name or a full commit ID).

$sel:destinationCommitSpecifier:CreateUnreferencedMergeCommit', createUnreferencedMergeCommit_destinationCommitSpecifier - The branch, tag, HEAD, or other fully qualified reference used to identify a commit (for example, a branch name or a full commit ID).

$sel:mergeOption:CreateUnreferencedMergeCommit', createUnreferencedMergeCommit_mergeOption - The merge option or strategy you want to use to merge the code.

Request Lenses

createUnreferencedMergeCommit_email :: Lens' CreateUnreferencedMergeCommit (Maybe Text)

The email address for the person who created the unreferenced commit.

createUnreferencedMergeCommit_authorName :: Lens' CreateUnreferencedMergeCommit (Maybe Text)

The name of the author who created the unreferenced commit. This information is used as both the author and committer for the commit.

createUnreferencedMergeCommit_conflictDetailLevel :: Lens' CreateUnreferencedMergeCommit (Maybe ConflictDetailLevelTypeEnum)

The level of conflict detail to use. If unspecified, the default FILE_LEVEL is used, which returns a not-mergeable result if the same file has differences in both branches. If LINE_LEVEL is specified, a conflict is considered not mergeable if the same file in both branches has differences on the same line.

createUnreferencedMergeCommit_commitMessage :: Lens' CreateUnreferencedMergeCommit (Maybe Text)

The commit message for the unreferenced commit.

createUnreferencedMergeCommit_conflictResolution :: Lens' CreateUnreferencedMergeCommit (Maybe ConflictResolution)

If AUTOMERGE is the conflict resolution strategy, a list of inputs to use when resolving conflicts during a merge.

createUnreferencedMergeCommit_conflictResolutionStrategy :: Lens' CreateUnreferencedMergeCommit (Maybe ConflictResolutionStrategyTypeEnum)

Specifies which branch to use when resolving conflicts, or whether to attempt automatically merging two versions of a file. The default is NONE, which requires any conflicts to be resolved manually before the merge operation is successful.

createUnreferencedMergeCommit_keepEmptyFolders :: Lens' CreateUnreferencedMergeCommit (Maybe Bool)

If the commit contains deletions, whether to keep a folder or folder structure if the changes leave the folders empty. If this is specified as true, a .gitkeep file is created for empty folders. The default is false.

createUnreferencedMergeCommit_repositoryName :: Lens' CreateUnreferencedMergeCommit Text

The name of the repository where you want to create the unreferenced merge commit.

createUnreferencedMergeCommit_sourceCommitSpecifier :: Lens' CreateUnreferencedMergeCommit Text

The branch, tag, HEAD, or other fully qualified reference used to identify a commit (for example, a branch name or a full commit ID).

createUnreferencedMergeCommit_destinationCommitSpecifier :: Lens' CreateUnreferencedMergeCommit Text

The branch, tag, HEAD, or other fully qualified reference used to identify a commit (for example, a branch name or a full commit ID).

createUnreferencedMergeCommit_mergeOption :: Lens' CreateUnreferencedMergeCommit MergeOptionTypeEnum

The merge option or strategy you want to use to merge the code.

Destructuring the Response

data CreateUnreferencedMergeCommitResponse

See: newCreateUnreferencedMergeCommitResponse smart constructor.

Constructors

CreateUnreferencedMergeCommitResponse'

Fields

commitId :: Maybe Text

The full commit ID of the commit that contains your merge results.

treeId :: Maybe Text

The full SHA-1 pointer of the tree information for the commit that contains the merge results.

httpStatus :: Int

The response's http status code.

newCreateUnreferencedMergeCommitResponse

Arguments

:: Int
$sel:httpStatus:CreateUnreferencedMergeCommitResponse'

-> CreateUnreferencedMergeCommitResponse

Create a value of CreateUnreferencedMergeCommitResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:commitId:CreateUnreferencedMergeCommitResponse', createUnreferencedMergeCommitResponse_commitId - The full commit ID of the commit that contains your merge results.

$sel:treeId:CreateUnreferencedMergeCommitResponse', createUnreferencedMergeCommitResponse_treeId - The full SHA-1 pointer of the tree information for the commit that contains the merge results.

$sel:httpStatus:CreateUnreferencedMergeCommitResponse', createUnreferencedMergeCommitResponse_httpStatus - The response's http status code.

Response Lenses

createUnreferencedMergeCommitResponse_commitId :: Lens' CreateUnreferencedMergeCommitResponse (Maybe Text)

The full commit ID of the commit that contains your merge results.

createUnreferencedMergeCommitResponse_treeId :: Lens' CreateUnreferencedMergeCommitResponse (Maybe Text)

The full SHA-1 pointer of the tree information for the commit that contains the merge results.

createUnreferencedMergeCommitResponse_httpStatus :: Lens' CreateUnreferencedMergeCommitResponse Int

The response's http status code.