Amazonka.CodeCommit.GetPullRequestApprovalStates

Description

Gets information about the approval states for a specified pull request. Approval states only apply to pull requests that have one or more approval rules applied to them.

Creating a Request

data GetPullRequestApprovalStates

See: newGetPullRequestApprovalStates smart constructor.

Constructors

GetPullRequestApprovalStates'

Fields

pullRequestId :: Text

The system-generated ID for the pull request.

revisionId :: Text

The system-generated ID for the pull request revision.

newGetPullRequestApprovalStates

Arguments

:: Text
$sel:pullRequestId:GetPullRequestApprovalStates'

-> Text
$sel:revisionId:GetPullRequestApprovalStates'

-> GetPullRequestApprovalStates

Create a value of GetPullRequestApprovalStates with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:pullRequestId:GetPullRequestApprovalStates', getPullRequestApprovalStates_pullRequestId - The system-generated ID for the pull request.

$sel:revisionId:GetPullRequestApprovalStates', getPullRequestApprovalStates_revisionId - The system-generated ID for the pull request revision.

Request Lenses

getPullRequestApprovalStates_pullRequestId :: Lens' GetPullRequestApprovalStates Text

The system-generated ID for the pull request.

getPullRequestApprovalStates_revisionId :: Lens' GetPullRequestApprovalStates Text

The system-generated ID for the pull request revision.

Destructuring the Response

data GetPullRequestApprovalStatesResponse

See: newGetPullRequestApprovalStatesResponse smart constructor.

Constructors

GetPullRequestApprovalStatesResponse'

Fields

approvals :: Maybe [Approval]

Information about users who have approved the pull request.

httpStatus :: Int

The response's http status code.

newGetPullRequestApprovalStatesResponse

Arguments

:: Int
$sel:httpStatus:GetPullRequestApprovalStatesResponse'

-> GetPullRequestApprovalStatesResponse

Create a value of GetPullRequestApprovalStatesResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:approvals:GetPullRequestApprovalStatesResponse', getPullRequestApprovalStatesResponse_approvals - Information about users who have approved the pull request.

$sel:httpStatus:GetPullRequestApprovalStatesResponse', getPullRequestApprovalStatesResponse_httpStatus - The response's http status code.

Response Lenses

getPullRequestApprovalStatesResponse_approvals :: Lens' GetPullRequestApprovalStatesResponse (Maybe [Approval])

Information about users who have approved the pull request.

getPullRequestApprovalStatesResponse_httpStatus :: Lens' GetPullRequestApprovalStatesResponse Int

The response's http status code.