Amazonka.CodeCommit.ListAssociatedApprovalRuleTemplatesForRepository

Description

Lists all approval rule templates that are associated with a specified repository.

Creating a Request

data ListAssociatedApprovalRuleTemplatesForRepository

See: newListAssociatedApprovalRuleTemplatesForRepository smart constructor.

Constructors

ListAssociatedApprovalRuleTemplatesForRepository'

Fields

nextToken :: Maybe Text

An enumeration token that, when provided in a request, returns the next batch of the results.

maxResults :: Maybe Int

A non-zero, non-negative integer used to limit the number of returned results.

repositoryName :: Text

The name of the repository for which you want to list all associated approval rule templates.

newListAssociatedApprovalRuleTemplatesForRepository

Arguments

:: Text
$sel:repositoryName:ListAssociatedApprovalRuleTemplatesForRepository'

-> ListAssociatedApprovalRuleTemplatesForRepository

Create a value of ListAssociatedApprovalRuleTemplatesForRepository with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:nextToken:ListAssociatedApprovalRuleTemplatesForRepository', listAssociatedApprovalRuleTemplatesForRepository_nextToken - An enumeration token that, when provided in a request, returns the next batch of the results.

$sel:maxResults:ListAssociatedApprovalRuleTemplatesForRepository', listAssociatedApprovalRuleTemplatesForRepository_maxResults - A non-zero, non-negative integer used to limit the number of returned results.

$sel:repositoryName:ListAssociatedApprovalRuleTemplatesForRepository', listAssociatedApprovalRuleTemplatesForRepository_repositoryName - The name of the repository for which you want to list all associated approval rule templates.

Request Lenses

listAssociatedApprovalRuleTemplatesForRepository_nextToken :: Lens' ListAssociatedApprovalRuleTemplatesForRepository (Maybe Text)

An enumeration token that, when provided in a request, returns the next batch of the results.

listAssociatedApprovalRuleTemplatesForRepository_maxResults :: Lens' ListAssociatedApprovalRuleTemplatesForRepository (Maybe Int)

A non-zero, non-negative integer used to limit the number of returned results.

listAssociatedApprovalRuleTemplatesForRepository_repositoryName :: Lens' ListAssociatedApprovalRuleTemplatesForRepository Text

The name of the repository for which you want to list all associated approval rule templates.

Destructuring the Response

data ListAssociatedApprovalRuleTemplatesForRepositoryResponse

See: newListAssociatedApprovalRuleTemplatesForRepositoryResponse smart constructor.

Constructors

ListAssociatedApprovalRuleTemplatesForRepositoryResponse'

Fields

nextToken :: Maybe Text

An enumeration token that allows the operation to batch the next results of the operation.

approvalRuleTemplateNames :: Maybe [Text]

The names of all approval rule templates associated with the repository.

httpStatus :: Int

The response's http status code.

newListAssociatedApprovalRuleTemplatesForRepositoryResponse

Arguments

:: Int
$sel:httpStatus:ListAssociatedApprovalRuleTemplatesForRepositoryResponse'

-> ListAssociatedApprovalRuleTemplatesForRepositoryResponse

Create a value of ListAssociatedApprovalRuleTemplatesForRepositoryResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:nextToken:ListAssociatedApprovalRuleTemplatesForRepository', listAssociatedApprovalRuleTemplatesForRepositoryResponse_nextToken - An enumeration token that allows the operation to batch the next results of the operation.

$sel:approvalRuleTemplateNames:ListAssociatedApprovalRuleTemplatesForRepositoryResponse', listAssociatedApprovalRuleTemplatesForRepositoryResponse_approvalRuleTemplateNames - The names of all approval rule templates associated with the repository.

$sel:httpStatus:ListAssociatedApprovalRuleTemplatesForRepositoryResponse', listAssociatedApprovalRuleTemplatesForRepositoryResponse_httpStatus - The response's http status code.

Response Lenses

listAssociatedApprovalRuleTemplatesForRepositoryResponse_nextToken :: Lens' ListAssociatedApprovalRuleTemplatesForRepositoryResponse (Maybe Text)

An enumeration token that allows the operation to batch the next results of the operation.

listAssociatedApprovalRuleTemplatesForRepositoryResponse_approvalRuleTemplateNames :: Lens' ListAssociatedApprovalRuleTemplatesForRepositoryResponse (Maybe [Text])

The names of all approval rule templates associated with the repository.

listAssociatedApprovalRuleTemplatesForRepositoryResponse_httpStatus :: Lens' ListAssociatedApprovalRuleTemplatesForRepositoryResponse Int

The response's http status code.