Amazonka.CodeCommit.ListBranches

Description

Gets information about one or more branches in a repository.

This operation returns paginated results.

Creating a Request

data ListBranches

Represents the input of a list branches operation.

See: newListBranches smart constructor.

Constructors

ListBranches'

Fields

nextToken :: Maybe Text

An enumeration token that allows the operation to batch the results.

repositoryName :: Text

The name of the repository that contains the branches.

newListBranches

Arguments

:: Text
$sel:repositoryName:ListBranches'

-> ListBranches

Create a value of ListBranches with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:nextToken:ListBranches', listBranches_nextToken - An enumeration token that allows the operation to batch the results.

$sel:repositoryName:ListBranches', listBranches_repositoryName - The name of the repository that contains the branches.

Request Lenses

listBranches_nextToken :: Lens' ListBranches (Maybe Text)

An enumeration token that allows the operation to batch the results.

listBranches_repositoryName :: Lens' ListBranches Text

The name of the repository that contains the branches.

Destructuring the Response

data ListBranchesResponse

Represents the output of a list branches operation.

See: newListBranchesResponse smart constructor.

Constructors

ListBranchesResponse'

Fields

branches :: Maybe [Text]

The list of branch names.

nextToken :: Maybe Text

An enumeration token that returns the batch of the results.

httpStatus :: Int

The response's http status code.

newListBranchesResponse

Arguments

:: Int
$sel:httpStatus:ListBranchesResponse'

-> ListBranchesResponse

Create a value of ListBranchesResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:branches:ListBranchesResponse', listBranchesResponse_branches - The list of branch names.

$sel:nextToken:ListBranches', listBranchesResponse_nextToken - An enumeration token that returns the batch of the results.

$sel:httpStatus:ListBranchesResponse', listBranchesResponse_httpStatus - The response's http status code.

Response Lenses

listBranchesResponse_branches :: Lens' ListBranchesResponse (Maybe [Text])

The list of branch names.

listBranchesResponse_nextToken :: Lens' ListBranchesResponse (Maybe Text)

An enumeration token that returns the batch of the results.

listBranchesResponse_httpStatus :: Lens' ListBranchesResponse Int

The response's http status code.