Amazonka.CodeCommit.MergePullRequestBySquash

Description

Attempts to merge the source commit of a pull request into the specified destination branch for that pull request at the specified commit using the squash merge strategy. If the merge is successful, it closes the pull request.

Creating a Request

data MergePullRequestBySquash

See: newMergePullRequestBySquash smart constructor.

Constructors

MergePullRequestBySquash'

Fields

email :: Maybe Text

The email address of the person merging the branches. This information is used in the commit information for the merge.

authorName :: Maybe Text

The name of the author who created the commit. This information is used as both the author and committer for the commit.

conflictDetailLevel :: Maybe ConflictDetailLevelTypeEnum

The level of conflict detail to use. If unspecified, the default FILE_LEVEL is used, which returns a not-mergeable result if the same file has differences in both branches. If LINE_LEVEL is specified, a conflict is considered not mergeable if the same file in both branches has differences on the same line.

commitMessage :: Maybe Text

The commit message to include in the commit information for the merge.

conflictResolution :: Maybe ConflictResolution

If AUTOMERGE is the conflict resolution strategy, a list of inputs to use when resolving conflicts during a merge.

conflictResolutionStrategy :: Maybe ConflictResolutionStrategyTypeEnum

Specifies which branch to use when resolving conflicts, or whether to attempt automatically merging two versions of a file. The default is NONE, which requires any conflicts to be resolved manually before the merge operation is successful.

keepEmptyFolders :: Maybe Bool

If the commit contains deletions, whether to keep a folder or folder structure if the changes leave the folders empty. If true, a .gitkeep file is created for empty folders. The default is false.

sourceCommitId :: Maybe Text

The full commit ID of the original or updated commit in the pull request source branch. Pass this value if you want an exception thrown if the current commit ID of the tip of the source branch does not match this commit ID.

pullRequestId :: Text

The system-generated ID of the pull request. To get this ID, use ListPullRequests.

repositoryName :: Text

The name of the repository where the pull request was created.

newMergePullRequestBySquash

Arguments

:: Text
$sel:pullRequestId:MergePullRequestBySquash'

-> Text
$sel:repositoryName:MergePullRequestBySquash'

-> MergePullRequestBySquash

Create a value of MergePullRequestBySquash with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:email:MergePullRequestBySquash', mergePullRequestBySquash_email - The email address of the person merging the branches. This information is used in the commit information for the merge.

$sel:authorName:MergePullRequestBySquash', mergePullRequestBySquash_authorName - The name of the author who created the commit. This information is used as both the author and committer for the commit.

$sel:conflictDetailLevel:MergePullRequestBySquash', mergePullRequestBySquash_conflictDetailLevel - The level of conflict detail to use. If unspecified, the default FILE_LEVEL is used, which returns a not-mergeable result if the same file has differences in both branches. If LINE_LEVEL is specified, a conflict is considered not mergeable if the same file in both branches has differences on the same line.

$sel:commitMessage:MergePullRequestBySquash', mergePullRequestBySquash_commitMessage - The commit message to include in the commit information for the merge.

$sel:conflictResolution:MergePullRequestBySquash', mergePullRequestBySquash_conflictResolution - If AUTOMERGE is the conflict resolution strategy, a list of inputs to use when resolving conflicts during a merge.

$sel:conflictResolutionStrategy:MergePullRequestBySquash', mergePullRequestBySquash_conflictResolutionStrategy - Specifies which branch to use when resolving conflicts, or whether to attempt automatically merging two versions of a file. The default is NONE, which requires any conflicts to be resolved manually before the merge operation is successful.

$sel:keepEmptyFolders:MergePullRequestBySquash', mergePullRequestBySquash_keepEmptyFolders - If the commit contains deletions, whether to keep a folder or folder structure if the changes leave the folders empty. If true, a .gitkeep file is created for empty folders. The default is false.

$sel:sourceCommitId:MergePullRequestBySquash', mergePullRequestBySquash_sourceCommitId - The full commit ID of the original or updated commit in the pull request source branch. Pass this value if you want an exception thrown if the current commit ID of the tip of the source branch does not match this commit ID.

$sel:pullRequestId:MergePullRequestBySquash', mergePullRequestBySquash_pullRequestId - The system-generated ID of the pull request. To get this ID, use ListPullRequests.

$sel:repositoryName:MergePullRequestBySquash', mergePullRequestBySquash_repositoryName - The name of the repository where the pull request was created.

Request Lenses

mergePullRequestBySquash_email :: Lens' MergePullRequestBySquash (Maybe Text)

The email address of the person merging the branches. This information is used in the commit information for the merge.

mergePullRequestBySquash_authorName :: Lens' MergePullRequestBySquash (Maybe Text)

The name of the author who created the commit. This information is used as both the author and committer for the commit.

mergePullRequestBySquash_conflictDetailLevel :: Lens' MergePullRequestBySquash (Maybe ConflictDetailLevelTypeEnum)

The level of conflict detail to use. If unspecified, the default FILE_LEVEL is used, which returns a not-mergeable result if the same file has differences in both branches. If LINE_LEVEL is specified, a conflict is considered not mergeable if the same file in both branches has differences on the same line.

mergePullRequestBySquash_commitMessage :: Lens' MergePullRequestBySquash (Maybe Text)

The commit message to include in the commit information for the merge.

mergePullRequestBySquash_conflictResolution :: Lens' MergePullRequestBySquash (Maybe ConflictResolution)

If AUTOMERGE is the conflict resolution strategy, a list of inputs to use when resolving conflicts during a merge.

mergePullRequestBySquash_conflictResolutionStrategy :: Lens' MergePullRequestBySquash (Maybe ConflictResolutionStrategyTypeEnum)

Specifies which branch to use when resolving conflicts, or whether to attempt automatically merging two versions of a file. The default is NONE, which requires any conflicts to be resolved manually before the merge operation is successful.

mergePullRequestBySquash_keepEmptyFolders :: Lens' MergePullRequestBySquash (Maybe Bool)

If the commit contains deletions, whether to keep a folder or folder structure if the changes leave the folders empty. If true, a .gitkeep file is created for empty folders. The default is false.

mergePullRequestBySquash_sourceCommitId :: Lens' MergePullRequestBySquash (Maybe Text)

The full commit ID of the original or updated commit in the pull request source branch. Pass this value if you want an exception thrown if the current commit ID of the tip of the source branch does not match this commit ID.

mergePullRequestBySquash_pullRequestId :: Lens' MergePullRequestBySquash Text

The system-generated ID of the pull request. To get this ID, use ListPullRequests.

mergePullRequestBySquash_repositoryName :: Lens' MergePullRequestBySquash Text

The name of the repository where the pull request was created.

Destructuring the Response

data MergePullRequestBySquashResponse

See: newMergePullRequestBySquashResponse smart constructor.

Constructors

MergePullRequestBySquashResponse'

Fields

pullRequest :: Maybe PullRequest

httpStatus :: Int

The response's http status code.

newMergePullRequestBySquashResponse

Arguments

:: Int
$sel:httpStatus:MergePullRequestBySquashResponse'

-> MergePullRequestBySquashResponse

Create a value of MergePullRequestBySquashResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:pullRequest:MergePullRequestBySquashResponse', mergePullRequestBySquashResponse_pullRequest - Undocumented member.

$sel:httpStatus:MergePullRequestBySquashResponse', mergePullRequestBySquashResponse_httpStatus - The response's http status code.

Response Lenses

mergePullRequestBySquashResponse_pullRequest :: Lens' MergePullRequestBySquashResponse (Maybe PullRequest)

Undocumented member.

mergePullRequestBySquashResponse_httpStatus :: Lens' MergePullRequestBySquashResponse Int

The response's http status code.