Amazonka.CodeCommit.PostCommentForPullRequest

Description

Posts a comment on a pull request.

Creating a Request

data PostCommentForPullRequest

See: newPostCommentForPullRequest smart constructor.

Constructors

PostCommentForPullRequest'

Fields

location :: Maybe Location

The location of the change where you want to post your comment. If no location is provided, the comment is posted as a general comment on the pull request difference between the before commit ID and the after commit ID.

clientRequestToken :: Maybe Text

A unique, client-generated idempotency token that, when provided in a request, ensures the request cannot be repeated with a changed parameter. If a request is received with the same parameters and a token is included, the request returns information about the initial request that used that token.

pullRequestId :: Text

The system-generated ID of the pull request. To get this ID, use ListPullRequests.

repositoryName :: Text

The name of the repository where you want to post a comment on a pull request.

beforeCommitId :: Text

The full commit ID of the commit in the destination branch that was the tip of the branch at the time the pull request was created.

afterCommitId :: Text

The full commit ID of the commit in the source branch that is the current tip of the branch for the pull request when you post the comment.

content :: Text

The content of your comment on the change.

newPostCommentForPullRequest

Arguments

:: Text
$sel:pullRequestId:PostCommentForPullRequest'

-> Text
$sel:repositoryName:PostCommentForPullRequest'

-> Text
$sel:beforeCommitId:PostCommentForPullRequest'

-> Text
$sel:afterCommitId:PostCommentForPullRequest'

-> Text
$sel:content:PostCommentForPullRequest'

-> PostCommentForPullRequest

Create a value of PostCommentForPullRequest with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:location:PostCommentForPullRequest', postCommentForPullRequest_location - The location of the change where you want to post your comment. If no location is provided, the comment is posted as a general comment on the pull request difference between the before commit ID and the after commit ID.

$sel:clientRequestToken:PostCommentForPullRequest', postCommentForPullRequest_clientRequestToken - A unique, client-generated idempotency token that, when provided in a request, ensures the request cannot be repeated with a changed parameter. If a request is received with the same parameters and a token is included, the request returns information about the initial request that used that token.

$sel:pullRequestId:PostCommentForPullRequest', postCommentForPullRequest_pullRequestId - The system-generated ID of the pull request. To get this ID, use ListPullRequests.

$sel:repositoryName:PostCommentForPullRequest', postCommentForPullRequest_repositoryName - The name of the repository where you want to post a comment on a pull request.

$sel:beforeCommitId:PostCommentForPullRequest', postCommentForPullRequest_beforeCommitId - The full commit ID of the commit in the destination branch that was the tip of the branch at the time the pull request was created.

$sel:afterCommitId:PostCommentForPullRequest', postCommentForPullRequest_afterCommitId - The full commit ID of the commit in the source branch that is the current tip of the branch for the pull request when you post the comment.

$sel:content:PostCommentForPullRequest', postCommentForPullRequest_content - The content of your comment on the change.

Request Lenses

postCommentForPullRequest_location :: Lens' PostCommentForPullRequest (Maybe Location)

The location of the change where you want to post your comment. If no location is provided, the comment is posted as a general comment on the pull request difference between the before commit ID and the after commit ID.

postCommentForPullRequest_clientRequestToken :: Lens' PostCommentForPullRequest (Maybe Text)

A unique, client-generated idempotency token that, when provided in a request, ensures the request cannot be repeated with a changed parameter. If a request is received with the same parameters and a token is included, the request returns information about the initial request that used that token.

postCommentForPullRequest_pullRequestId :: Lens' PostCommentForPullRequest Text

The system-generated ID of the pull request. To get this ID, use ListPullRequests.

postCommentForPullRequest_repositoryName :: Lens' PostCommentForPullRequest Text

The name of the repository where you want to post a comment on a pull request.

postCommentForPullRequest_beforeCommitId :: Lens' PostCommentForPullRequest Text

The full commit ID of the commit in the destination branch that was the tip of the branch at the time the pull request was created.

postCommentForPullRequest_afterCommitId :: Lens' PostCommentForPullRequest Text

The full commit ID of the commit in the source branch that is the current tip of the branch for the pull request when you post the comment.

postCommentForPullRequest_content :: Lens' PostCommentForPullRequest Text

The content of your comment on the change.

Destructuring the Response

data PostCommentForPullRequestResponse

See: newPostCommentForPullRequestResponse smart constructor.

Constructors

PostCommentForPullRequestResponse'

Fields

beforeBlobId :: Maybe Text

In the directionality of the pull request, the blob ID of the before blob.

location :: Maybe Location

The location of the change where you posted your comment.

afterCommitId :: Maybe Text

The full commit ID of the commit in the destination branch where the pull request is merged.

pullRequestId :: Maybe Text

The system-generated ID of the pull request.

afterBlobId :: Maybe Text

In the directionality of the pull request, the blob ID of the after blob.

beforeCommitId :: Maybe Text

The full commit ID of the commit in the source branch used to create the pull request, or in the case of an updated pull request, the full commit ID of the commit used to update the pull request.

repositoryName :: Maybe Text

The name of the repository where you posted a comment on a pull request.

comment :: Maybe Comment

The content of the comment you posted.

httpStatus :: Int

The response's http status code.

newPostCommentForPullRequestResponse

Arguments

:: Int
$sel:httpStatus:PostCommentForPullRequestResponse'

-> PostCommentForPullRequestResponse

Create a value of PostCommentForPullRequestResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:beforeBlobId:PostCommentForPullRequestResponse', postCommentForPullRequestResponse_beforeBlobId - In the directionality of the pull request, the blob ID of the before blob.

$sel:location:PostCommentForPullRequest', postCommentForPullRequestResponse_location - The location of the change where you posted your comment.

$sel:afterCommitId:PostCommentForPullRequest', postCommentForPullRequestResponse_afterCommitId - The full commit ID of the commit in the destination branch where the pull request is merged.

$sel:pullRequestId:PostCommentForPullRequest', postCommentForPullRequestResponse_pullRequestId - The system-generated ID of the pull request.

$sel:afterBlobId:PostCommentForPullRequestResponse', postCommentForPullRequestResponse_afterBlobId - In the directionality of the pull request, the blob ID of the after blob.

$sel:beforeCommitId:PostCommentForPullRequest', postCommentForPullRequestResponse_beforeCommitId - The full commit ID of the commit in the source branch used to create the pull request, or in the case of an updated pull request, the full commit ID of the commit used to update the pull request.

$sel:repositoryName:PostCommentForPullRequest', postCommentForPullRequestResponse_repositoryName - The name of the repository where you posted a comment on a pull request.

$sel:comment:PostCommentForPullRequestResponse', postCommentForPullRequestResponse_comment - The content of the comment you posted.

$sel:httpStatus:PostCommentForPullRequestResponse', postCommentForPullRequestResponse_httpStatus - The response's http status code.

Response Lenses

postCommentForPullRequestResponse_beforeBlobId :: Lens' PostCommentForPullRequestResponse (Maybe Text)

In the directionality of the pull request, the blob ID of the before blob.

postCommentForPullRequestResponse_location :: Lens' PostCommentForPullRequestResponse (Maybe Location)

The location of the change where you posted your comment.

postCommentForPullRequestResponse_afterCommitId :: Lens' PostCommentForPullRequestResponse (Maybe Text)

The full commit ID of the commit in the destination branch where the pull request is merged.

postCommentForPullRequestResponse_pullRequestId :: Lens' PostCommentForPullRequestResponse (Maybe Text)

The system-generated ID of the pull request.

postCommentForPullRequestResponse_afterBlobId :: Lens' PostCommentForPullRequestResponse (Maybe Text)

In the directionality of the pull request, the blob ID of the after blob.

postCommentForPullRequestResponse_beforeCommitId :: Lens' PostCommentForPullRequestResponse (Maybe Text)

The full commit ID of the commit in the source branch used to create the pull request, or in the case of an updated pull request, the full commit ID of the commit used to update the pull request.

postCommentForPullRequestResponse_repositoryName :: Lens' PostCommentForPullRequestResponse (Maybe Text)

The name of the repository where you posted a comment on a pull request.

postCommentForPullRequestResponse_comment :: Lens' PostCommentForPullRequestResponse (Maybe Comment)

The content of the comment you posted.

postCommentForPullRequestResponse_httpStatus :: Lens' PostCommentForPullRequestResponse Int

The response's http status code.