Amazonka.CodeCommit.Types.ApprovalStateChangedEventMetadata

Description

Documentation

data ApprovalStateChangedEventMetadata

Returns information about a change in the approval state for a pull request.

See: newApprovalStateChangedEventMetadata smart constructor.

Constructors

ApprovalStateChangedEventMetadata'

Fields

approvalStatus :: Maybe ApprovalState

The approval status for the pull request.

revisionId :: Maybe Text

The revision ID of the pull request when the approval state changed.

newApprovalStateChangedEventMetadata :: ApprovalStateChangedEventMetadata

Create a value of ApprovalStateChangedEventMetadata with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:approvalStatus:ApprovalStateChangedEventMetadata', approvalStateChangedEventMetadata_approvalStatus - The approval status for the pull request.

$sel:revisionId:ApprovalStateChangedEventMetadata', approvalStateChangedEventMetadata_revisionId - The revision ID of the pull request when the approval state changed.

approvalStateChangedEventMetadata_approvalStatus :: Lens' ApprovalStateChangedEventMetadata (Maybe ApprovalState)

The approval status for the pull request.

approvalStateChangedEventMetadata_revisionId :: Lens' ApprovalStateChangedEventMetadata (Maybe Text)

The revision ID of the pull request when the approval state changed.