Amazonka.CodeCommit.Types.Comment

Description

Documentation

data Comment

Returns information about a specific comment.

See: newComment smart constructor.

Constructors

Comment'

Fields

lastModifiedDate :: Maybe POSIX

The date and time the comment was most recently modified, in timestamp format.

authorArn :: Maybe Text

The Amazon Resource Name (ARN) of the person who posted the comment.

content :: Maybe Text

The content of the comment.

callerReactions :: Maybe [Text]

The emoji reactions to a comment, if any, submitted by the user whose credentials are associated with the call to the API.

creationDate :: Maybe POSIX

The date and time the comment was created, in timestamp format.

deleted :: Maybe Bool

A Boolean value indicating whether the comment has been deleted.

clientRequestToken :: Maybe Text

A unique, client-generated idempotency token that, when provided in a request, ensures the request cannot be repeated with a changed parameter. If a request is received with the same parameters and a token is included, the request returns information about the initial request that used that token.

commentId :: Maybe Text

The system-generated comment ID.

inReplyTo :: Maybe Text

The ID of the comment for which this comment is a reply, if any.

reactionCounts :: Maybe (HashMap Text Int)

A string to integer map that represents the number of individual users who have responded to a comment with the specified reactions.

newComment :: Comment

Create a value of Comment with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:lastModifiedDate:Comment', comment_lastModifiedDate - The date and time the comment was most recently modified, in timestamp format.

$sel:authorArn:Comment', comment_authorArn - The Amazon Resource Name (ARN) of the person who posted the comment.

$sel:content:Comment', comment_content - The content of the comment.

$sel:callerReactions:Comment', comment_callerReactions - The emoji reactions to a comment, if any, submitted by the user whose credentials are associated with the call to the API.

$sel:creationDate:Comment', comment_creationDate - The date and time the comment was created, in timestamp format.

$sel:deleted:Comment', comment_deleted - A Boolean value indicating whether the comment has been deleted.

$sel:clientRequestToken:Comment', comment_clientRequestToken - A unique, client-generated idempotency token that, when provided in a request, ensures the request cannot be repeated with a changed parameter. If a request is received with the same parameters and a token is included, the request returns information about the initial request that used that token.

$sel:commentId:Comment', comment_commentId - The system-generated comment ID.

$sel:inReplyTo:Comment', comment_inReplyTo - The ID of the comment for which this comment is a reply, if any.

$sel:reactionCounts:Comment', comment_reactionCounts - A string to integer map that represents the number of individual users who have responded to a comment with the specified reactions.

comment_lastModifiedDate :: Lens' Comment (Maybe UTCTime)

The date and time the comment was most recently modified, in timestamp format.

comment_authorArn :: Lens' Comment (Maybe Text)

The Amazon Resource Name (ARN) of the person who posted the comment.

comment_content :: Lens' Comment (Maybe Text)

The content of the comment.

comment_callerReactions :: Lens' Comment (Maybe [Text])

The emoji reactions to a comment, if any, submitted by the user whose credentials are associated with the call to the API.

comment_creationDate :: Lens' Comment (Maybe UTCTime)

The date and time the comment was created, in timestamp format.

comment_deleted :: Lens' Comment (Maybe Bool)

A Boolean value indicating whether the comment has been deleted.

comment_clientRequestToken :: Lens' Comment (Maybe Text)

A unique, client-generated idempotency token that, when provided in a request, ensures the request cannot be repeated with a changed parameter. If a request is received with the same parameters and a token is included, the request returns information about the initial request that used that token.

comment_commentId :: Lens' Comment (Maybe Text)

The system-generated comment ID.

comment_inReplyTo :: Lens' Comment (Maybe Text)

The ID of the comment for which this comment is a reply, if any.

comment_reactionCounts :: Lens' Comment (Maybe (HashMap Text Int))

A string to integer map that represents the number of individual users who have responded to a comment with the specified reactions.