Amazonka.CodeCommit.Types.CommentsForComparedCommit

Description

Documentation

data CommentsForComparedCommit

Returns information about comments on the comparison between two commits.

See: newCommentsForComparedCommit smart constructor.

Constructors

CommentsForComparedCommit'

Fields

beforeBlobId :: Maybe Text

The full blob ID of the commit used to establish the before of the comparison.

location :: Maybe Location

Location information about the comment on the comparison, including the file name, line number, and whether the version of the file where the comment was made is BEFORE or AFTER.

afterCommitId :: Maybe Text

The full commit ID of the commit used to establish the after of the comparison.

afterBlobId :: Maybe Text

The full blob ID of the commit used to establish the after of the comparison.

beforeCommitId :: Maybe Text

The full commit ID of the commit used to establish the before of the comparison.

repositoryName :: Maybe Text

The name of the repository that contains the compared commits.

comments :: Maybe [Comment]

An array of comment objects. Each comment object contains information about a comment on the comparison between commits.

newCommentsForComparedCommit :: CommentsForComparedCommit

Create a value of CommentsForComparedCommit with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:beforeBlobId:CommentsForComparedCommit', commentsForComparedCommit_beforeBlobId - The full blob ID of the commit used to establish the before of the comparison.

$sel:location:CommentsForComparedCommit', commentsForComparedCommit_location - Location information about the comment on the comparison, including the file name, line number, and whether the version of the file where the comment was made is BEFORE or AFTER.

$sel:afterCommitId:CommentsForComparedCommit', commentsForComparedCommit_afterCommitId - The full commit ID of the commit used to establish the after of the comparison.

$sel:afterBlobId:CommentsForComparedCommit', commentsForComparedCommit_afterBlobId - The full blob ID of the commit used to establish the after of the comparison.

$sel:beforeCommitId:CommentsForComparedCommit', commentsForComparedCommit_beforeCommitId - The full commit ID of the commit used to establish the before of the comparison.

$sel:repositoryName:CommentsForComparedCommit', commentsForComparedCommit_repositoryName - The name of the repository that contains the compared commits.

$sel:comments:CommentsForComparedCommit', commentsForComparedCommit_comments - An array of comment objects. Each comment object contains information about a comment on the comparison between commits.

commentsForComparedCommit_beforeBlobId :: Lens' CommentsForComparedCommit (Maybe Text)

The full blob ID of the commit used to establish the before of the comparison.

commentsForComparedCommit_location :: Lens' CommentsForComparedCommit (Maybe Location)

Location information about the comment on the comparison, including the file name, line number, and whether the version of the file where the comment was made is BEFORE or AFTER.

commentsForComparedCommit_afterCommitId :: Lens' CommentsForComparedCommit (Maybe Text)

The full commit ID of the commit used to establish the after of the comparison.

commentsForComparedCommit_afterBlobId :: Lens' CommentsForComparedCommit (Maybe Text)

The full blob ID of the commit used to establish the after of the comparison.

commentsForComparedCommit_beforeCommitId :: Lens' CommentsForComparedCommit (Maybe Text)

The full commit ID of the commit used to establish the before of the comparison.

commentsForComparedCommit_repositoryName :: Lens' CommentsForComparedCommit (Maybe Text)

The name of the repository that contains the compared commits.

commentsForComparedCommit_comments :: Lens' CommentsForComparedCommit (Maybe [Comment])

An array of comment objects. Each comment object contains information about a comment on the comparison between commits.