Amazonka.CodeCommit.Types.File

Description

Documentation

data File

Returns information about a file in a repository.

See: newFile smart constructor.

Constructors

File'

Fields

absolutePath :: Maybe Text

The fully qualified path to the file in the repository.

fileMode :: Maybe FileModeTypeEnum

The extrapolated file mode permissions for the file. Valid values include EXECUTABLE and NORMAL.

blobId :: Maybe Text

The blob ID that contains the file information.

relativePath :: Maybe Text

The relative path of the file from the folder where the query originated.

newFile :: File

Create a value of File with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:absolutePath:File', file_absolutePath - The fully qualified path to the file in the repository.

$sel:fileMode:File', file_fileMode - The extrapolated file mode permissions for the file. Valid values include EXECUTABLE and NORMAL.

$sel:blobId:File', file_blobId - The blob ID that contains the file information.

$sel:relativePath:File', file_relativePath - The relative path of the file from the folder where the query originated.

file_absolutePath :: Lens' File (Maybe Text)

The fully qualified path to the file in the repository.

file_fileMode :: Lens' File (Maybe FileModeTypeEnum)

The extrapolated file mode permissions for the file. Valid values include EXECUTABLE and NORMAL.

file_blobId :: Lens' File (Maybe Text)

The blob ID that contains the file information.

file_relativePath :: Lens' File (Maybe Text)

The relative path of the file from the folder where the query originated.