Amazonka.CodeDeploy.CreateDeployment

Description

Deploys an application revision through the specified deployment group.

Creating a Request

data CreateDeployment

Represents the input of a CreateDeployment operation.

See: newCreateDeployment smart constructor.

Constructors

CreateDeployment'

Fields

deploymentConfigName :: Maybe Text

The name of a deployment configuration associated with the IAM user or AWS account.

If not specified, the value configured in the deployment group is used as the default. If the deployment group does not have a deployment configuration associated with it, CodeDeployDefault.OneAtATime is used by default.

fileExistsBehavior :: Maybe FileExistsBehavior

Information about how AWS CodeDeploy handles files that already exist in a deployment target location but weren't part of the previous successful deployment.

The fileExistsBehavior parameter takes any of the following values:

DISALLOW: The deployment fails. This is also the default behavior if no option is specified.

OVERWRITE: The version of the file from the application revision currently being deployed replaces the version already on the instance.

RETAIN: The version of the file already on the instance is kept and used as part of the new deployment.

targetInstances :: Maybe TargetInstances

Information about the instances that belong to the replacement environment in a blue/green deployment.

revision :: Maybe RevisionLocation

The type and location of the revision to deploy.

description :: Maybe Text

A comment about the deployment.

autoRollbackConfiguration :: Maybe AutoRollbackConfiguration

Configuration information for an automatic rollback that is added when a deployment is created.

updateOutdatedInstancesOnly :: Maybe Bool

Indicates whether to deploy to all instances or only to instances that are not running the latest application revision.

deploymentGroupName :: Maybe Text

The name of the deployment group.

ignoreApplicationStopFailures :: Maybe Bool

If true, then if an ApplicationStop, BeforeBlockTraffic, or AfterBlockTraffic deployment lifecycle event to an instance fails, then the deployment continues to the next deployment lifecycle event. For example, if ApplicationStop fails, the deployment continues with DownloadBundle. If BeforeBlockTraffic fails, the deployment continues with BlockTraffic. If AfterBlockTraffic fails, the deployment continues with ApplicationStop.

If false or not specified, then if a lifecycle event fails during a deployment to an instance, that deployment fails. If deployment to that instance is part of an overall deployment and the number of healthy hosts is not less than the minimum number of healthy hosts, then a deployment to the next instance is attempted.

During a deployment, the AWS CodeDeploy agent runs the scripts specified for ApplicationStop, BeforeBlockTraffic, and AfterBlockTraffic in the AppSpec file from the previous successful deployment. (All other scripts are run from the AppSpec file in the current deployment.) If one of these scripts contains an error and does not run successfully, the deployment can fail.

If the cause of the failure is a script from the last successful deployment that will never run successfully, create a new deployment and use ignoreApplicationStopFailures to specify that the ApplicationStop, BeforeBlockTraffic, and AfterBlockTraffic failures should be ignored.

applicationName :: Text

The name of an AWS CodeDeploy application associated with the IAM user or AWS account.

newCreateDeployment

Arguments

:: Text
$sel:applicationName:CreateDeployment'

-> CreateDeployment

Create a value of CreateDeployment with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:deploymentConfigName:CreateDeployment', createDeployment_deploymentConfigName - The name of a deployment configuration associated with the IAM user or AWS account.

If not specified, the value configured in the deployment group is used as the default. If the deployment group does not have a deployment configuration associated with it, CodeDeployDefault.OneAtATime is used by default.

$sel:fileExistsBehavior:CreateDeployment', createDeployment_fileExistsBehavior - Information about how AWS CodeDeploy handles files that already exist in a deployment target location but weren't part of the previous successful deployment.

The fileExistsBehavior parameter takes any of the following values:

DISALLOW: The deployment fails. This is also the default behavior if no option is specified.

OVERWRITE: The version of the file from the application revision currently being deployed replaces the version already on the instance.

RETAIN: The version of the file already on the instance is kept and used as part of the new deployment.

$sel:targetInstances:CreateDeployment', createDeployment_targetInstances - Information about the instances that belong to the replacement environment in a blue/green deployment.

$sel:revision:CreateDeployment', createDeployment_revision - The type and location of the revision to deploy.

$sel:description:CreateDeployment', createDeployment_description - A comment about the deployment.

$sel:autoRollbackConfiguration:CreateDeployment', createDeployment_autoRollbackConfiguration - Configuration information for an automatic rollback that is added when a deployment is created.

$sel:updateOutdatedInstancesOnly:CreateDeployment', createDeployment_updateOutdatedInstancesOnly - Indicates whether to deploy to all instances or only to instances that are not running the latest application revision.

$sel:deploymentGroupName:CreateDeployment', createDeployment_deploymentGroupName - The name of the deployment group.

$sel:ignoreApplicationStopFailures:CreateDeployment', createDeployment_ignoreApplicationStopFailures - If true, then if an ApplicationStop, BeforeBlockTraffic, or AfterBlockTraffic deployment lifecycle event to an instance fails, then the deployment continues to the next deployment lifecycle event. For example, if ApplicationStop fails, the deployment continues with DownloadBundle. If BeforeBlockTraffic fails, the deployment continues with BlockTraffic. If AfterBlockTraffic fails, the deployment continues with ApplicationStop.

If false or not specified, then if a lifecycle event fails during a deployment to an instance, that deployment fails. If deployment to that instance is part of an overall deployment and the number of healthy hosts is not less than the minimum number of healthy hosts, then a deployment to the next instance is attempted.

During a deployment, the AWS CodeDeploy agent runs the scripts specified for ApplicationStop, BeforeBlockTraffic, and AfterBlockTraffic in the AppSpec file from the previous successful deployment. (All other scripts are run from the AppSpec file in the current deployment.) If one of these scripts contains an error and does not run successfully, the deployment can fail.

If the cause of the failure is a script from the last successful deployment that will never run successfully, create a new deployment and use ignoreApplicationStopFailures to specify that the ApplicationStop, BeforeBlockTraffic, and AfterBlockTraffic failures should be ignored.

$sel:applicationName:CreateDeployment', createDeployment_applicationName - The name of an AWS CodeDeploy application associated with the IAM user or AWS account.

Request Lenses

createDeployment_deploymentConfigName :: Lens' CreateDeployment (Maybe Text)

The name of a deployment configuration associated with the IAM user or AWS account.

If not specified, the value configured in the deployment group is used as the default. If the deployment group does not have a deployment configuration associated with it, CodeDeployDefault.OneAtATime is used by default.

createDeployment_fileExistsBehavior :: Lens' CreateDeployment (Maybe FileExistsBehavior)

Information about how AWS CodeDeploy handles files that already exist in a deployment target location but weren't part of the previous successful deployment.

The fileExistsBehavior parameter takes any of the following values:

DISALLOW: The deployment fails. This is also the default behavior if no option is specified.

OVERWRITE: The version of the file from the application revision currently being deployed replaces the version already on the instance.

RETAIN: The version of the file already on the instance is kept and used as part of the new deployment.

createDeployment_targetInstances :: Lens' CreateDeployment (Maybe TargetInstances)

Information about the instances that belong to the replacement environment in a blue/green deployment.

createDeployment_revision :: Lens' CreateDeployment (Maybe RevisionLocation)

The type and location of the revision to deploy.

createDeployment_description :: Lens' CreateDeployment (Maybe Text)

A comment about the deployment.

createDeployment_autoRollbackConfiguration :: Lens' CreateDeployment (Maybe AutoRollbackConfiguration)

Configuration information for an automatic rollback that is added when a deployment is created.

createDeployment_updateOutdatedInstancesOnly :: Lens' CreateDeployment (Maybe Bool)

Indicates whether to deploy to all instances or only to instances that are not running the latest application revision.

createDeployment_deploymentGroupName :: Lens' CreateDeployment (Maybe Text)

The name of the deployment group.

createDeployment_ignoreApplicationStopFailures :: Lens' CreateDeployment (Maybe Bool)

If true, then if an ApplicationStop, BeforeBlockTraffic, or AfterBlockTraffic deployment lifecycle event to an instance fails, then the deployment continues to the next deployment lifecycle event. For example, if ApplicationStop fails, the deployment continues with DownloadBundle. If BeforeBlockTraffic fails, the deployment continues with BlockTraffic. If AfterBlockTraffic fails, the deployment continues with ApplicationStop.

If false or not specified, then if a lifecycle event fails during a deployment to an instance, that deployment fails. If deployment to that instance is part of an overall deployment and the number of healthy hosts is not less than the minimum number of healthy hosts, then a deployment to the next instance is attempted.

During a deployment, the AWS CodeDeploy agent runs the scripts specified for ApplicationStop, BeforeBlockTraffic, and AfterBlockTraffic in the AppSpec file from the previous successful deployment. (All other scripts are run from the AppSpec file in the current deployment.) If one of these scripts contains an error and does not run successfully, the deployment can fail.

If the cause of the failure is a script from the last successful deployment that will never run successfully, create a new deployment and use ignoreApplicationStopFailures to specify that the ApplicationStop, BeforeBlockTraffic, and AfterBlockTraffic failures should be ignored.

createDeployment_applicationName :: Lens' CreateDeployment Text

The name of an AWS CodeDeploy application associated with the IAM user or AWS account.

Destructuring the Response

data CreateDeploymentResponse

Represents the output of a CreateDeployment operation.

See: newCreateDeploymentResponse smart constructor.

Constructors

CreateDeploymentResponse'

Fields

deploymentId :: Maybe Text

The unique ID of a deployment.

httpStatus :: Int

The response's http status code.

newCreateDeploymentResponse

Arguments

:: Int
$sel:httpStatus:CreateDeploymentResponse'

-> CreateDeploymentResponse

Create a value of CreateDeploymentResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:deploymentId:CreateDeploymentResponse', createDeploymentResponse_deploymentId - The unique ID of a deployment.

$sel:httpStatus:CreateDeploymentResponse', createDeploymentResponse_httpStatus - The response's http status code.

Response Lenses

createDeploymentResponse_deploymentId :: Lens' CreateDeploymentResponse (Maybe Text)

The unique ID of a deployment.

createDeploymentResponse_httpStatus :: Lens' CreateDeploymentResponse Int

The response's http status code.