Amazonka.CodeDeploy.CreateDeploymentConfig

Description

Creates a deployment configuration.

Creating a Request

data CreateDeploymentConfig

Represents the input of a CreateDeploymentConfig operation.

See: newCreateDeploymentConfig smart constructor.

Constructors

CreateDeploymentConfig'

Fields

computePlatform :: Maybe ComputePlatform

The destination platform type for the deployment (Lambda, Server, or ECS).

minimumHealthyHosts :: Maybe MinimumHealthyHosts

The minimum number of healthy instances that should be available at any time during the deployment. There are two parameters expected in the input: type and value.

The type parameter takes either of the following values:

HOST_COUNT: The value parameter represents the minimum number of healthy instances as an absolute value.

FLEET_PERCENT: The value parameter represents the minimum number of healthy instances as a percentage of the total number of instances in the deployment. If you specify FLEET_PERCENT, at the start of the deployment, AWS CodeDeploy converts the percentage to the equivalent number of instances and rounds up fractional instances.

The value parameter takes an integer.

For example, to set a minimum of 95% healthy instance, specify a type of FLEET_PERCENT and a value of 95.

trafficRoutingConfig :: Maybe TrafficRoutingConfig

The configuration that specifies how the deployment traffic is routed.

deploymentConfigName :: Text

The name of the deployment configuration to create.

newCreateDeploymentConfig

Arguments

:: Text
$sel:deploymentConfigName:CreateDeploymentConfig'

-> CreateDeploymentConfig

Create a value of CreateDeploymentConfig with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:computePlatform:CreateDeploymentConfig', createDeploymentConfig_computePlatform - The destination platform type for the deployment (Lambda, Server, or ECS).

$sel:minimumHealthyHosts:CreateDeploymentConfig', createDeploymentConfig_minimumHealthyHosts - The minimum number of healthy instances that should be available at any time during the deployment. There are two parameters expected in the input: type and value.

The type parameter takes either of the following values:

HOST_COUNT: The value parameter represents the minimum number of healthy instances as an absolute value.

FLEET_PERCENT: The value parameter represents the minimum number of healthy instances as a percentage of the total number of instances in the deployment. If you specify FLEET_PERCENT, at the start of the deployment, AWS CodeDeploy converts the percentage to the equivalent number of instances and rounds up fractional instances.

The value parameter takes an integer.

For example, to set a minimum of 95% healthy instance, specify a type of FLEET_PERCENT and a value of 95.

$sel:trafficRoutingConfig:CreateDeploymentConfig', createDeploymentConfig_trafficRoutingConfig - The configuration that specifies how the deployment traffic is routed.

$sel:deploymentConfigName:CreateDeploymentConfig', createDeploymentConfig_deploymentConfigName - The name of the deployment configuration to create.

Request Lenses

createDeploymentConfig_computePlatform :: Lens' CreateDeploymentConfig (Maybe ComputePlatform)

The destination platform type for the deployment (Lambda, Server, or ECS).

createDeploymentConfig_minimumHealthyHosts :: Lens' CreateDeploymentConfig (Maybe MinimumHealthyHosts)

The minimum number of healthy instances that should be available at any time during the deployment. There are two parameters expected in the input: type and value.

The type parameter takes either of the following values:

HOST_COUNT: The value parameter represents the minimum number of healthy instances as an absolute value.

FLEET_PERCENT: The value parameter represents the minimum number of healthy instances as a percentage of the total number of instances in the deployment. If you specify FLEET_PERCENT, at the start of the deployment, AWS CodeDeploy converts the percentage to the equivalent number of instances and rounds up fractional instances.

The value parameter takes an integer.

For example, to set a minimum of 95% healthy instance, specify a type of FLEET_PERCENT and a value of 95.

createDeploymentConfig_trafficRoutingConfig :: Lens' CreateDeploymentConfig (Maybe TrafficRoutingConfig)

The configuration that specifies how the deployment traffic is routed.

createDeploymentConfig_deploymentConfigName :: Lens' CreateDeploymentConfig Text

The name of the deployment configuration to create.

Destructuring the Response

data CreateDeploymentConfigResponse

Represents the output of a CreateDeploymentConfig operation.

See: newCreateDeploymentConfigResponse smart constructor.

Constructors

CreateDeploymentConfigResponse'

Fields

deploymentConfigId :: Maybe Text

A unique deployment configuration ID.

httpStatus :: Int

The response's http status code.

newCreateDeploymentConfigResponse

Arguments

:: Int
$sel:httpStatus:CreateDeploymentConfigResponse'

-> CreateDeploymentConfigResponse

Create a value of CreateDeploymentConfigResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:deploymentConfigId:CreateDeploymentConfigResponse', createDeploymentConfigResponse_deploymentConfigId - A unique deployment configuration ID.

$sel:httpStatus:CreateDeploymentConfigResponse', createDeploymentConfigResponse_httpStatus - The response's http status code.

Response Lenses

createDeploymentConfigResponse_deploymentConfigId :: Lens' CreateDeploymentConfigResponse (Maybe Text)

A unique deployment configuration ID.

createDeploymentConfigResponse_httpStatus :: Lens' CreateDeploymentConfigResponse Int

The response's http status code.