Amazonka.CodeDeploy.GetDeploymentConfig

Description

Gets information about a deployment configuration.

Creating a Request

data GetDeploymentConfig

Represents the input of a GetDeploymentConfig operation.

See: newGetDeploymentConfig smart constructor.

Constructors

GetDeploymentConfig'

Fields

deploymentConfigName :: Text

The name of a deployment configuration associated with the IAM user or AWS account.

newGetDeploymentConfig

Arguments

:: Text
$sel:deploymentConfigName:GetDeploymentConfig'

-> GetDeploymentConfig

Create a value of GetDeploymentConfig with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:deploymentConfigName:GetDeploymentConfig', getDeploymentConfig_deploymentConfigName - The name of a deployment configuration associated with the IAM user or AWS account.

Request Lenses

getDeploymentConfig_deploymentConfigName :: Lens' GetDeploymentConfig Text

The name of a deployment configuration associated with the IAM user or AWS account.

Destructuring the Response

data GetDeploymentConfigResponse

Represents the output of a GetDeploymentConfig operation.

See: newGetDeploymentConfigResponse smart constructor.

Constructors

GetDeploymentConfigResponse'

Fields

deploymentConfigInfo :: Maybe DeploymentConfigInfo

Information about the deployment configuration.

httpStatus :: Int

The response's http status code.

newGetDeploymentConfigResponse

Arguments

:: Int
$sel:httpStatus:GetDeploymentConfigResponse'

-> GetDeploymentConfigResponse

Create a value of GetDeploymentConfigResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:deploymentConfigInfo:GetDeploymentConfigResponse', getDeploymentConfigResponse_deploymentConfigInfo - Information about the deployment configuration.

$sel:httpStatus:GetDeploymentConfigResponse', getDeploymentConfigResponse_httpStatus - The response's http status code.

Response Lenses

getDeploymentConfigResponse_deploymentConfigInfo :: Lens' GetDeploymentConfigResponse (Maybe DeploymentConfigInfo)

Information about the deployment configuration.

getDeploymentConfigResponse_httpStatus :: Lens' GetDeploymentConfigResponse Int

The response's http status code.