Amazonka.CodeDeploy.ListApplicationRevisions

Description

Lists information about revisions for an application.

This operation returns paginated results.

Creating a Request

data ListApplicationRevisions

Represents the input of a ListApplicationRevisions operation.

See: newListApplicationRevisions smart constructor.

Constructors

ListApplicationRevisions'

Fields

s3KeyPrefix :: Maybe Text

A key prefix for the set of Amazon S3 objects to limit the search for revisions.

deployed :: Maybe ListStateFilterAction

Whether to list revisions based on whether the revision is the target revision of a deployment group:

include: List revisions that are target revisions of a deployment group.

exclude: Do not list revisions that are target revisions of a deployment group.

ignore: List all revisions.

sortOrder :: Maybe SortOrder

The order in which to sort the list results:

ascending: ascending order.

descending: descending order.

If not specified, the results are sorted in ascending order.

If set to null, the results are sorted in an arbitrary order.

nextToken :: Maybe Text

An identifier returned from the previous ListApplicationRevisions call. It can be used to return the next set of applications in the list.

s3Bucket :: Maybe Text

An Amazon S3 bucket name to limit the search for revisions.

If set to null, all of the user's buckets are searched.

sortBy :: Maybe ApplicationRevisionSortBy

The column name to use to sort the list results:

registerTime: Sort by the time the revisions were registered with AWS CodeDeploy.

firstUsedTime: Sort by the time the revisions were first used in a deployment.

lastUsedTime: Sort by the time the revisions were last used in a deployment.

If not specified or set to null, the results are returned in an arbitrary order.

applicationName :: Text

The name of an AWS CodeDeploy application associated with the IAM user or AWS account.

newListApplicationRevisions

Arguments

:: Text
$sel:applicationName:ListApplicationRevisions'

-> ListApplicationRevisions

Create a value of ListApplicationRevisions with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:s3KeyPrefix:ListApplicationRevisions', listApplicationRevisions_s3KeyPrefix - A key prefix for the set of Amazon S3 objects to limit the search for revisions.

$sel:deployed:ListApplicationRevisions', listApplicationRevisions_deployed - Whether to list revisions based on whether the revision is the target revision of a deployment group:

include: List revisions that are target revisions of a deployment group.

exclude: Do not list revisions that are target revisions of a deployment group.

ignore: List all revisions.

$sel:sortOrder:ListApplicationRevisions', listApplicationRevisions_sortOrder - The order in which to sort the list results:

ascending: ascending order.

descending: descending order.

If not specified, the results are sorted in ascending order.

If set to null, the results are sorted in an arbitrary order.

$sel:nextToken:ListApplicationRevisions', listApplicationRevisions_nextToken - An identifier returned from the previous ListApplicationRevisions call. It can be used to return the next set of applications in the list.

$sel:s3Bucket:ListApplicationRevisions', listApplicationRevisions_s3Bucket - An Amazon S3 bucket name to limit the search for revisions.

If set to null, all of the user's buckets are searched.

$sel:sortBy:ListApplicationRevisions', listApplicationRevisions_sortBy - The column name to use to sort the list results:

registerTime: Sort by the time the revisions were registered with AWS CodeDeploy.

firstUsedTime: Sort by the time the revisions were first used in a deployment.

lastUsedTime: Sort by the time the revisions were last used in a deployment.

If not specified or set to null, the results are returned in an arbitrary order.

$sel:applicationName:ListApplicationRevisions', listApplicationRevisions_applicationName - The name of an AWS CodeDeploy application associated with the IAM user or AWS account.

Request Lenses

listApplicationRevisions_s3KeyPrefix :: Lens' ListApplicationRevisions (Maybe Text)

A key prefix for the set of Amazon S3 objects to limit the search for revisions.

listApplicationRevisions_deployed :: Lens' ListApplicationRevisions (Maybe ListStateFilterAction)

Whether to list revisions based on whether the revision is the target revision of a deployment group:

include: List revisions that are target revisions of a deployment group.

exclude: Do not list revisions that are target revisions of a deployment group.

ignore: List all revisions.

listApplicationRevisions_sortOrder :: Lens' ListApplicationRevisions (Maybe SortOrder)

The order in which to sort the list results:

ascending: ascending order.

descending: descending order.

If not specified, the results are sorted in ascending order.

If set to null, the results are sorted in an arbitrary order.

listApplicationRevisions_nextToken :: Lens' ListApplicationRevisions (Maybe Text)

An identifier returned from the previous ListApplicationRevisions call. It can be used to return the next set of applications in the list.

listApplicationRevisions_s3Bucket :: Lens' ListApplicationRevisions (Maybe Text)

An Amazon S3 bucket name to limit the search for revisions.

If set to null, all of the user's buckets are searched.

listApplicationRevisions_sortBy :: Lens' ListApplicationRevisions (Maybe ApplicationRevisionSortBy)

The column name to use to sort the list results:

registerTime: Sort by the time the revisions were registered with AWS CodeDeploy.

firstUsedTime: Sort by the time the revisions were first used in a deployment.

lastUsedTime: Sort by the time the revisions were last used in a deployment.

If not specified or set to null, the results are returned in an arbitrary order.

listApplicationRevisions_applicationName :: Lens' ListApplicationRevisions Text

The name of an AWS CodeDeploy application associated with the IAM user or AWS account.

Destructuring the Response

data ListApplicationRevisionsResponse

Represents the output of a ListApplicationRevisions operation.

See: newListApplicationRevisionsResponse smart constructor.

Constructors

ListApplicationRevisionsResponse'

Fields

nextToken :: Maybe Text

If a large amount of information is returned, an identifier is also returned. It can be used in a subsequent list application revisions call to return the next set of application revisions in the list.

revisions :: Maybe [RevisionLocation]

A list of locations that contain the matching revisions.

httpStatus :: Int

The response's http status code.

newListApplicationRevisionsResponse

Arguments

:: Int
$sel:httpStatus:ListApplicationRevisionsResponse'

-> ListApplicationRevisionsResponse

Create a value of ListApplicationRevisionsResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:nextToken:ListApplicationRevisions', listApplicationRevisionsResponse_nextToken - If a large amount of information is returned, an identifier is also returned. It can be used in a subsequent list application revisions call to return the next set of application revisions in the list.

$sel:revisions:ListApplicationRevisionsResponse', listApplicationRevisionsResponse_revisions - A list of locations that contain the matching revisions.

$sel:httpStatus:ListApplicationRevisionsResponse', listApplicationRevisionsResponse_httpStatus - The response's http status code.

Response Lenses

listApplicationRevisionsResponse_nextToken :: Lens' ListApplicationRevisionsResponse (Maybe Text)

If a large amount of information is returned, an identifier is also returned. It can be used in a subsequent list application revisions call to return the next set of application revisions in the list.

listApplicationRevisionsResponse_revisions :: Lens' ListApplicationRevisionsResponse (Maybe [RevisionLocation])

A list of locations that contain the matching revisions.

listApplicationRevisionsResponse_httpStatus :: Lens' ListApplicationRevisionsResponse Int

The response's http status code.