Amazonka.CodeDeploy.Types.AutoRollbackConfiguration

Description

Documentation

data AutoRollbackConfiguration

Information about a configuration for automatically rolling back to a previous version of an application revision when a deployment is not completed successfully.

See: newAutoRollbackConfiguration smart constructor.

Constructors

AutoRollbackConfiguration'

Fields

enabled :: Maybe Bool

Indicates whether a defined automatic rollback configuration is currently enabled.

events :: Maybe [AutoRollbackEvent]

The event type or types that trigger a rollback.

newAutoRollbackConfiguration :: AutoRollbackConfiguration

Create a value of AutoRollbackConfiguration with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:enabled:AutoRollbackConfiguration', autoRollbackConfiguration_enabled - Indicates whether a defined automatic rollback configuration is currently enabled.

$sel:events:AutoRollbackConfiguration', autoRollbackConfiguration_events - The event type or types that trigger a rollback.

autoRollbackConfiguration_enabled :: Lens' AutoRollbackConfiguration (Maybe Bool)

Indicates whether a defined automatic rollback configuration is currently enabled.

autoRollbackConfiguration_events :: Lens' AutoRollbackConfiguration (Maybe [AutoRollbackEvent])

The event type or types that trigger a rollback.