Amazonka.CodeDeploy.Types.CloudFormationTarget

Description

Documentation

data CloudFormationTarget

Information about the target to be updated by an AWS CloudFormation blue/green deployment. This target type is used for all deployments initiated by a CloudFormation stack update.

See: newCloudFormationTarget smart constructor.

Constructors

CloudFormationTarget'

Fields

targetId :: Maybe Text

The unique ID of a deployment target that has a type of CloudFormationTarget.

status :: Maybe TargetStatus

The status of an AWS CloudFormation blue/green deployment's target application.

deploymentId :: Maybe Text

The unique ID of an AWS CloudFormation blue/green deployment.

resourceType :: Maybe Text

The resource type for the AWS CloudFormation blue/green deployment.

lastUpdatedAt :: Maybe POSIX

The date and time when the target application was updated by an AWS CloudFormation blue/green deployment.

lifecycleEvents :: Maybe [LifecycleEvent]

The lifecycle events of the AWS CloudFormation blue/green deployment to this target application.

targetVersionWeight :: Maybe Double

The percentage of production traffic that the target version of an AWS CloudFormation blue/green deployment receives.

newCloudFormationTarget :: CloudFormationTarget

Create a value of CloudFormationTarget with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:targetId:CloudFormationTarget', cloudFormationTarget_targetId - The unique ID of a deployment target that has a type of CloudFormationTarget.

$sel:status:CloudFormationTarget', cloudFormationTarget_status - The status of an AWS CloudFormation blue/green deployment's target application.

$sel:deploymentId:CloudFormationTarget', cloudFormationTarget_deploymentId - The unique ID of an AWS CloudFormation blue/green deployment.

$sel:resourceType:CloudFormationTarget', cloudFormationTarget_resourceType - The resource type for the AWS CloudFormation blue/green deployment.

$sel:lastUpdatedAt:CloudFormationTarget', cloudFormationTarget_lastUpdatedAt - The date and time when the target application was updated by an AWS CloudFormation blue/green deployment.

$sel:lifecycleEvents:CloudFormationTarget', cloudFormationTarget_lifecycleEvents - The lifecycle events of the AWS CloudFormation blue/green deployment to this target application.

$sel:targetVersionWeight:CloudFormationTarget', cloudFormationTarget_targetVersionWeight - The percentage of production traffic that the target version of an AWS CloudFormation blue/green deployment receives.

cloudFormationTarget_targetId :: Lens' CloudFormationTarget (Maybe Text)

The unique ID of a deployment target that has a type of CloudFormationTarget.

cloudFormationTarget_status :: Lens' CloudFormationTarget (Maybe TargetStatus)

The status of an AWS CloudFormation blue/green deployment's target application.

cloudFormationTarget_deploymentId :: Lens' CloudFormationTarget (Maybe Text)

The unique ID of an AWS CloudFormation blue/green deployment.

cloudFormationTarget_resourceType :: Lens' CloudFormationTarget (Maybe Text)

The resource type for the AWS CloudFormation blue/green deployment.

cloudFormationTarget_lastUpdatedAt :: Lens' CloudFormationTarget (Maybe UTCTime)

The date and time when the target application was updated by an AWS CloudFormation blue/green deployment.

cloudFormationTarget_lifecycleEvents :: Lens' CloudFormationTarget (Maybe [LifecycleEvent])

The lifecycle events of the AWS CloudFormation blue/green deployment to this target application.

cloudFormationTarget_targetVersionWeight :: Lens' CloudFormationTarget (Maybe Double)

The percentage of production traffic that the target version of an AWS CloudFormation blue/green deployment receives.