Amazonka.CodeDeploy.Types.DeploymentConfigInfo

Description

Documentation

data DeploymentConfigInfo

Information about a deployment configuration.

See: newDeploymentConfigInfo smart constructor.

Constructors

DeploymentConfigInfo'

Fields

deploymentConfigName :: Maybe Text

The deployment configuration name.

computePlatform :: Maybe ComputePlatform

The destination platform type for the deployment (Lambda, Server, or ECS).

minimumHealthyHosts :: Maybe MinimumHealthyHosts

Information about the number or percentage of minimum healthy instance.

trafficRoutingConfig :: Maybe TrafficRoutingConfig

The configuration that specifies how the deployment traffic is routed. Used for deployments with a Lambda or ECS compute platform only.

deploymentConfigId :: Maybe Text

The deployment configuration ID.

createTime :: Maybe POSIX

The time at which the deployment configuration was created.

newDeploymentConfigInfo :: DeploymentConfigInfo

Create a value of DeploymentConfigInfo with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:deploymentConfigName:DeploymentConfigInfo', deploymentConfigInfo_deploymentConfigName - The deployment configuration name.

$sel:computePlatform:DeploymentConfigInfo', deploymentConfigInfo_computePlatform - The destination platform type for the deployment (Lambda, Server, or ECS).

$sel:minimumHealthyHosts:DeploymentConfigInfo', deploymentConfigInfo_minimumHealthyHosts - Information about the number or percentage of minimum healthy instance.

$sel:trafficRoutingConfig:DeploymentConfigInfo', deploymentConfigInfo_trafficRoutingConfig - The configuration that specifies how the deployment traffic is routed. Used for deployments with a Lambda or ECS compute platform only.

$sel:deploymentConfigId:DeploymentConfigInfo', deploymentConfigInfo_deploymentConfigId - The deployment configuration ID.

$sel:createTime:DeploymentConfigInfo', deploymentConfigInfo_createTime - The time at which the deployment configuration was created.

deploymentConfigInfo_deploymentConfigName :: Lens' DeploymentConfigInfo (Maybe Text)

The deployment configuration name.

deploymentConfigInfo_computePlatform :: Lens' DeploymentConfigInfo (Maybe ComputePlatform)

The destination platform type for the deployment (Lambda, Server, or ECS).

deploymentConfigInfo_minimumHealthyHosts :: Lens' DeploymentConfigInfo (Maybe MinimumHealthyHosts)

Information about the number or percentage of minimum healthy instance.

deploymentConfigInfo_trafficRoutingConfig :: Lens' DeploymentConfigInfo (Maybe TrafficRoutingConfig)

The configuration that specifies how the deployment traffic is routed. Used for deployments with a Lambda or ECS compute platform only.

deploymentConfigInfo_deploymentConfigId :: Lens' DeploymentConfigInfo (Maybe Text)

The deployment configuration ID.

deploymentConfigInfo_createTime :: Lens' DeploymentConfigInfo (Maybe UTCTime)

The time at which the deployment configuration was created.