Amazonka.CodeDeploy.Types.DeploymentReadyOption

Description

Documentation

data DeploymentReadyOption

Information about how traffic is rerouted to instances in a replacement environment in a blue/green deployment.

See: newDeploymentReadyOption smart constructor.

Constructors

DeploymentReadyOption'

Fields

actionOnTimeout :: Maybe DeploymentReadyAction

Information about when to reroute traffic from an original environment to a replacement environment in a blue/green deployment.

CONTINUE_DEPLOYMENT: Register new instances with the load balancer immediately after the new application revision is installed on the instances in the replacement environment.

STOP_DEPLOYMENT: Do not register new instances with a load balancer unless traffic rerouting is started using ContinueDeployment. If traffic rerouting is not started before the end of the specified wait period, the deployment status is changed to Stopped.

waitTimeInMinutes :: Maybe Int

The number of minutes to wait before the status of a blue/green deployment is changed to Stopped if rerouting is not started manually. Applies only to the STOP_DEPLOYMENT option for actionOnTimeout.

newDeploymentReadyOption :: DeploymentReadyOption

Create a value of DeploymentReadyOption with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:actionOnTimeout:DeploymentReadyOption', deploymentReadyOption_actionOnTimeout - Information about when to reroute traffic from an original environment to a replacement environment in a blue/green deployment.

CONTINUE_DEPLOYMENT: Register new instances with the load balancer immediately after the new application revision is installed on the instances in the replacement environment.

STOP_DEPLOYMENT: Do not register new instances with a load balancer unless traffic rerouting is started using ContinueDeployment. If traffic rerouting is not started before the end of the specified wait period, the deployment status is changed to Stopped.

$sel:waitTimeInMinutes:DeploymentReadyOption', deploymentReadyOption_waitTimeInMinutes - The number of minutes to wait before the status of a blue/green deployment is changed to Stopped if rerouting is not started manually. Applies only to the STOP_DEPLOYMENT option for actionOnTimeout.

deploymentReadyOption_actionOnTimeout :: Lens' DeploymentReadyOption (Maybe DeploymentReadyAction)

Information about when to reroute traffic from an original environment to a replacement environment in a blue/green deployment.

CONTINUE_DEPLOYMENT: Register new instances with the load balancer immediately after the new application revision is installed on the instances in the replacement environment.

STOP_DEPLOYMENT: Do not register new instances with a load balancer unless traffic rerouting is started using ContinueDeployment. If traffic rerouting is not started before the end of the specified wait period, the deployment status is changed to Stopped.

deploymentReadyOption_waitTimeInMinutes :: Lens' DeploymentReadyOption (Maybe Int)

The number of minutes to wait before the status of a blue/green deployment is changed to Stopped if rerouting is not started manually. Applies only to the STOP_DEPLOYMENT option for actionOnTimeout.