Amazonka.CodeDeploy.Types.Diagnostics

Description

Documentation

data Diagnostics

Diagnostic information about executable scripts that are part of a deployment.

See: newDiagnostics smart constructor.

Constructors

Diagnostics'

Fields

logTail :: Maybe Text

The last portion of the diagnostic log.

If available, AWS CodeDeploy returns up to the last 4 KB of the diagnostic log.

errorCode :: Maybe LifecycleErrorCode

The associated error code:

Success: The specified script ran.

ScriptMissing: The specified script was not found in the specified location.

ScriptNotExecutable: The specified script is not a recognized executable file type.

ScriptTimedOut: The specified script did not finish running in the specified time period.

ScriptFailed: The specified script failed to run as expected.

UnknownError: The specified script did not run for an unknown reason.

scriptName :: Maybe Text

The name of the script.

message :: Maybe Text

The message associated with the error.

newDiagnostics :: Diagnostics

Create a value of Diagnostics with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:logTail:Diagnostics', diagnostics_logTail - The last portion of the diagnostic log.

If available, AWS CodeDeploy returns up to the last 4 KB of the diagnostic log.

$sel:errorCode:Diagnostics', diagnostics_errorCode - The associated error code:

Success: The specified script ran.

ScriptMissing: The specified script was not found in the specified location.

ScriptNotExecutable: The specified script is not a recognized executable file type.

ScriptTimedOut: The specified script did not finish running in the specified time period.

ScriptFailed: The specified script failed to run as expected.

UnknownError: The specified script did not run for an unknown reason.

$sel:scriptName:Diagnostics', diagnostics_scriptName - The name of the script.

$sel:message:Diagnostics', diagnostics_message - The message associated with the error.

diagnostics_logTail :: Lens' Diagnostics (Maybe Text)

The last portion of the diagnostic log.

If available, AWS CodeDeploy returns up to the last 4 KB of the diagnostic log.

diagnostics_errorCode :: Lens' Diagnostics (Maybe LifecycleErrorCode)

The associated error code:

Success: The specified script ran.

ScriptMissing: The specified script was not found in the specified location.

ScriptNotExecutable: The specified script is not a recognized executable file type.

ScriptTimedOut: The specified script did not finish running in the specified time period.

ScriptFailed: The specified script failed to run as expected.

UnknownError: The specified script did not run for an unknown reason.

diagnostics_scriptName :: Lens' Diagnostics (Maybe Text)

The name of the script.

diagnostics_message :: Lens' Diagnostics (Maybe Text)

The message associated with the error.