Amazonka.CodeDeploy.Types.EC2TagSet

Description

Documentation

data EC2TagSet

Information about groups of EC2 instance tags.

See: newEC2TagSet smart constructor.

Constructors

EC2TagSet'

Fields

ec2TagSetList :: Maybe [[EC2TagFilter]]

A list that contains other lists of EC2 instance tag groups. For an instance to be included in the deployment group, it must be identified by all of the tag groups in the list.

newEC2TagSet :: EC2TagSet

Create a value of EC2TagSet with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:ec2TagSetList:EC2TagSet', eC2TagSet_ec2TagSetList - A list that contains other lists of EC2 instance tag groups. For an instance to be included in the deployment group, it must be identified by all of the tag groups in the list.

eC2TagSet_ec2TagSetList :: Lens' EC2TagSet (Maybe [[EC2TagFilter]])

A list that contains other lists of EC2 instance tag groups. For an instance to be included in the deployment group, it must be identified by all of the tag groups in the list.