Amazonka.CodeDeploy.Types.ECSTarget

Description

Documentation

data ECSTarget

Information about the target of an Amazon ECS deployment.

See: newECSTarget smart constructor.

Constructors

ECSTarget'

Fields

targetArn :: Maybe Text

The Amazon Resource Name (ARN) of the target.

targetId :: Maybe Text

The unique ID of a deployment target that has a type of ecsTarget.

status :: Maybe TargetStatus

The status an Amazon ECS deployment's target ECS application.

deploymentId :: Maybe Text

The unique ID of a deployment.

lastUpdatedAt :: Maybe POSIX

The date and time when the target Amazon ECS application was updated by a deployment.

taskSetsInfo :: Maybe [ECSTaskSet]

The ECSTaskSet objects associated with the ECS target.

lifecycleEvents :: Maybe [LifecycleEvent]

The lifecycle events of the deployment to this target Amazon ECS application.

newECSTarget :: ECSTarget

Create a value of ECSTarget with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:targetArn:ECSTarget', eCSTarget_targetArn - The Amazon Resource Name (ARN) of the target.

$sel:targetId:ECSTarget', eCSTarget_targetId - The unique ID of a deployment target that has a type of ecsTarget.

$sel:status:ECSTarget', eCSTarget_status - The status an Amazon ECS deployment's target ECS application.

$sel:deploymentId:ECSTarget', eCSTarget_deploymentId - The unique ID of a deployment.

$sel:lastUpdatedAt:ECSTarget', eCSTarget_lastUpdatedAt - The date and time when the target Amazon ECS application was updated by a deployment.

$sel:taskSetsInfo:ECSTarget', eCSTarget_taskSetsInfo - The ECSTaskSet objects associated with the ECS target.

$sel:lifecycleEvents:ECSTarget', eCSTarget_lifecycleEvents - The lifecycle events of the deployment to this target Amazon ECS application.

eCSTarget_targetArn :: Lens' ECSTarget (Maybe Text)

The Amazon Resource Name (ARN) of the target.

eCSTarget_targetId :: Lens' ECSTarget (Maybe Text)

The unique ID of a deployment target that has a type of ecsTarget.

eCSTarget_status :: Lens' ECSTarget (Maybe TargetStatus)

The status an Amazon ECS deployment's target ECS application.

eCSTarget_deploymentId :: Lens' ECSTarget (Maybe Text)

The unique ID of a deployment.

eCSTarget_lastUpdatedAt :: Lens' ECSTarget (Maybe UTCTime)

The date and time when the target Amazon ECS application was updated by a deployment.

eCSTarget_taskSetsInfo :: Lens' ECSTarget (Maybe [ECSTaskSet])

The ECSTaskSet objects associated with the ECS target.

eCSTarget_lifecycleEvents :: Lens' ECSTarget (Maybe [LifecycleEvent])

The lifecycle events of the deployment to this target Amazon ECS application.