Amazonka.CodeDeploy.Types.LastDeploymentInfo

Description

Documentation

data LastDeploymentInfo

Information about the most recent attempted or successful deployment to a deployment group.

See: newLastDeploymentInfo smart constructor.

Constructors

LastDeploymentInfo'

Fields

status :: Maybe DeploymentStatus

The status of the most recent deployment.

deploymentId :: Maybe Text

The unique ID of a deployment.

endTime :: Maybe POSIX

A timestamp that indicates when the most recent deployment to the deployment group was complete.

createTime :: Maybe POSIX

A timestamp that indicates when the most recent deployment to the deployment group started.

newLastDeploymentInfo :: LastDeploymentInfo

Create a value of LastDeploymentInfo with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:status:LastDeploymentInfo', lastDeploymentInfo_status - The status of the most recent deployment.

$sel:deploymentId:LastDeploymentInfo', lastDeploymentInfo_deploymentId - The unique ID of a deployment.

$sel:endTime:LastDeploymentInfo', lastDeploymentInfo_endTime - A timestamp that indicates when the most recent deployment to the deployment group was complete.

$sel:createTime:LastDeploymentInfo', lastDeploymentInfo_createTime - A timestamp that indicates when the most recent deployment to the deployment group started.

lastDeploymentInfo_status :: Lens' LastDeploymentInfo (Maybe DeploymentStatus)

The status of the most recent deployment.

lastDeploymentInfo_deploymentId :: Lens' LastDeploymentInfo (Maybe Text)

The unique ID of a deployment.

lastDeploymentInfo_endTime :: Lens' LastDeploymentInfo (Maybe UTCTime)

A timestamp that indicates when the most recent deployment to the deployment group was complete.

lastDeploymentInfo_createTime :: Lens' LastDeploymentInfo (Maybe UTCTime)

A timestamp that indicates when the most recent deployment to the deployment group started.