Amazonka.CodeDeploy.Types.LoadBalancerInfo

Description

Documentation

data LoadBalancerInfo

Information about the Elastic Load Balancing load balancer or target group used in a deployment.

See: newLoadBalancerInfo smart constructor.

Constructors

LoadBalancerInfo'

Fields

elbInfoList :: Maybe [ELBInfo]

An array that contains information about the load balancer to use for load balancing in a deployment. In Elastic Load Balancing, load balancers are used with Classic Load Balancers.

Adding more than one load balancer to the array is not supported.

targetGroupInfoList :: Maybe [TargetGroupInfo]

An array that contains information about the target group to use for load balancing in a deployment. In Elastic Load Balancing, target groups are used with Application Load Balancers.

Adding more than one target group to the array is not supported.

targetGroupPairInfoList :: Maybe [TargetGroupPairInfo]

The target group pair information. This is an array of TargeGroupPairInfo objects with a maximum size of one.

newLoadBalancerInfo :: LoadBalancerInfo

Create a value of LoadBalancerInfo with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:elbInfoList:LoadBalancerInfo', loadBalancerInfo_elbInfoList - An array that contains information about the load balancer to use for load balancing in a deployment. In Elastic Load Balancing, load balancers are used with Classic Load Balancers.

Adding more than one load balancer to the array is not supported.

$sel:targetGroupInfoList:LoadBalancerInfo', loadBalancerInfo_targetGroupInfoList - An array that contains information about the target group to use for load balancing in a deployment. In Elastic Load Balancing, target groups are used with Application Load Balancers.

Adding more than one target group to the array is not supported.

$sel:targetGroupPairInfoList:LoadBalancerInfo', loadBalancerInfo_targetGroupPairInfoList - The target group pair information. This is an array of TargeGroupPairInfo objects with a maximum size of one.

loadBalancerInfo_elbInfoList :: Lens' LoadBalancerInfo (Maybe [ELBInfo])

An array that contains information about the load balancer to use for load balancing in a deployment. In Elastic Load Balancing, load balancers are used with Classic Load Balancers.

Adding more than one load balancer to the array is not supported.

loadBalancerInfo_targetGroupInfoList :: Lens' LoadBalancerInfo (Maybe [TargetGroupInfo])

An array that contains information about the target group to use for load balancing in a deployment. In Elastic Load Balancing, target groups are used with Application Load Balancers.

Adding more than one target group to the array is not supported.

loadBalancerInfo_targetGroupPairInfoList :: Lens' LoadBalancerInfo (Maybe [TargetGroupPairInfo])

The target group pair information. This is an array of TargeGroupPairInfo objects with a maximum size of one.