Amazonka.CodeGuruProfiler.GetFindingsReportAccountSummary

Description

Returns a list of FindingsReportSummary objects that contain analysis results for all profiling groups in your AWS account.

Creating a Request

data GetFindingsReportAccountSummary

The structure representing the GetFindingsReportAccountSummaryRequest.

See: newGetFindingsReportAccountSummary smart constructor.

Constructors

GetFindingsReportAccountSummary'

Fields

nextToken :: Maybe Text

The nextToken value returned from a previous paginated GetFindingsReportAccountSummary request where maxResults was used and the results exceeded the value of that parameter. Pagination continues from the end of the previous results that returned the nextToken value.

This token should be treated as an opaque identifier that is only used to retrieve the next items in a list and not for other programmatic purposes.

dailyReportsOnly :: Maybe Bool

A Boolean value indicating whether to only return reports from daily profiles. If set to True, only analysis data from daily profiles is returned. If set to False, analysis data is returned from smaller time windows (for example, one hour).

maxResults :: Maybe Natural

The maximum number of results returned by GetFindingsReportAccountSummary in paginated output. When this parameter is used, GetFindingsReportAccountSummary only returns maxResults results in a single page along with a nextToken response element. The remaining results of the initial request can be seen by sending another GetFindingsReportAccountSummary request with the returned nextToken value.

newGetFindingsReportAccountSummary :: GetFindingsReportAccountSummary

Create a value of GetFindingsReportAccountSummary with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:nextToken:GetFindingsReportAccountSummary', getFindingsReportAccountSummary_nextToken - The nextToken value returned from a previous paginated GetFindingsReportAccountSummary request where maxResults was used and the results exceeded the value of that parameter. Pagination continues from the end of the previous results that returned the nextToken value.

This token should be treated as an opaque identifier that is only used to retrieve the next items in a list and not for other programmatic purposes.

$sel:dailyReportsOnly:GetFindingsReportAccountSummary', getFindingsReportAccountSummary_dailyReportsOnly - A Boolean value indicating whether to only return reports from daily profiles. If set to True, only analysis data from daily profiles is returned. If set to False, analysis data is returned from smaller time windows (for example, one hour).

$sel:maxResults:GetFindingsReportAccountSummary', getFindingsReportAccountSummary_maxResults - The maximum number of results returned by GetFindingsReportAccountSummary in paginated output. When this parameter is used, GetFindingsReportAccountSummary only returns maxResults results in a single page along with a nextToken response element. The remaining results of the initial request can be seen by sending another GetFindingsReportAccountSummary request with the returned nextToken value.

Request Lenses

getFindingsReportAccountSummary_nextToken :: Lens' GetFindingsReportAccountSummary (Maybe Text)

The nextToken value returned from a previous paginated GetFindingsReportAccountSummary request where maxResults was used and the results exceeded the value of that parameter. Pagination continues from the end of the previous results that returned the nextToken value.

This token should be treated as an opaque identifier that is only used to retrieve the next items in a list and not for other programmatic purposes.

getFindingsReportAccountSummary_dailyReportsOnly :: Lens' GetFindingsReportAccountSummary (Maybe Bool)

A Boolean value indicating whether to only return reports from daily profiles. If set to True, only analysis data from daily profiles is returned. If set to False, analysis data is returned from smaller time windows (for example, one hour).

getFindingsReportAccountSummary_maxResults :: Lens' GetFindingsReportAccountSummary (Maybe Natural)

The maximum number of results returned by GetFindingsReportAccountSummary in paginated output. When this parameter is used, GetFindingsReportAccountSummary only returns maxResults results in a single page along with a nextToken response element. The remaining results of the initial request can be seen by sending another GetFindingsReportAccountSummary request with the returned nextToken value.

Destructuring the Response

data GetFindingsReportAccountSummaryResponse

The structure representing the GetFindingsReportAccountSummaryResponse.

See: newGetFindingsReportAccountSummaryResponse smart constructor.

Constructors

GetFindingsReportAccountSummaryResponse'

Fields

nextToken :: Maybe Text

The nextToken value to include in a future GetFindingsReportAccountSummary request. When the results of a GetFindingsReportAccountSummary request exceed maxResults, this value can be used to retrieve the next page of results. This value is null when there are no more results to return.

httpStatus :: Int

The response's http status code.

reportSummaries :: [FindingsReportSummary]

The return list of FindingsReportSummary objects taht contain summaries of analysis results for all profiling groups in your AWS account.

newGetFindingsReportAccountSummaryResponse

Arguments

:: Int
$sel:httpStatus:GetFindingsReportAccountSummaryResponse'

-> GetFindingsReportAccountSummaryResponse

Create a value of GetFindingsReportAccountSummaryResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:nextToken:GetFindingsReportAccountSummary', getFindingsReportAccountSummaryResponse_nextToken - The nextToken value to include in a future GetFindingsReportAccountSummary request. When the results of a GetFindingsReportAccountSummary request exceed maxResults, this value can be used to retrieve the next page of results. This value is null when there are no more results to return.

$sel:httpStatus:GetFindingsReportAccountSummaryResponse', getFindingsReportAccountSummaryResponse_httpStatus - The response's http status code.

$sel:reportSummaries:GetFindingsReportAccountSummaryResponse', getFindingsReportAccountSummaryResponse_reportSummaries - The return list of FindingsReportSummary objects taht contain summaries of analysis results for all profiling groups in your AWS account.

Response Lenses

getFindingsReportAccountSummaryResponse_nextToken :: Lens' GetFindingsReportAccountSummaryResponse (Maybe Text)

The nextToken value to include in a future GetFindingsReportAccountSummary request. When the results of a GetFindingsReportAccountSummary request exceed maxResults, this value can be used to retrieve the next page of results. This value is null when there are no more results to return.

getFindingsReportAccountSummaryResponse_httpStatus :: Lens' GetFindingsReportAccountSummaryResponse Int

The response's http status code.

getFindingsReportAccountSummaryResponse_reportSummaries :: Lens' GetFindingsReportAccountSummaryResponse [FindingsReportSummary]

The return list of FindingsReportSummary objects taht contain summaries of analysis results for all profiling groups in your AWS account.