Amazonka.CodeGuruProfiler.PostAgentProfile

Description

Submits profiling data to an aggregated profile of a profiling group. To get an aggregated profile that is created with this profiling data, use GetProfile .

Creating a Request

data PostAgentProfile

The structure representing the postAgentProfileRequest.

See: newPostAgentProfile smart constructor.

Constructors

PostAgentProfile'

Fields

profileToken :: Maybe Text

Amazon CodeGuru Profiler uses this universally unique identifier (UUID) to prevent the accidental submission of duplicate profiling data if there are failures and retries.

agentProfile :: ByteString

The submitted profiling data.

contentType :: Text

The format of the submitted profiling data. The format maps to the Accept and Content-Type headers of the HTTP request. You can specify one of the following: or the default .

 <ul> <li> <p> <code>application/json</code> — standard JSON format </p> </li> <li> <p> <code>application/x-amzn-ion</code> — the Amazon Ion data format. For more information, see <a href="http://amzn.github.io/ion-docs/">Amazon Ion</a>. </p> </li> </ul>

profilingGroupName :: Text

The name of the profiling group with the aggregated profile that receives the submitted profiling data.

newPostAgentProfile

Arguments

:: ByteString
$sel:agentProfile:PostAgentProfile'

-> Text
$sel:contentType:PostAgentProfile'

-> Text
$sel:profilingGroupName:PostAgentProfile'

-> PostAgentProfile

Create a value of PostAgentProfile with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:profileToken:PostAgentProfile', postAgentProfile_profileToken - Amazon CodeGuru Profiler uses this universally unique identifier (UUID) to prevent the accidental submission of duplicate profiling data if there are failures and retries.

$sel:agentProfile:PostAgentProfile', postAgentProfile_agentProfile - The submitted profiling data.

$sel:contentType:PostAgentProfile', postAgentProfile_contentType - The format of the submitted profiling data. The format maps to the Accept and Content-Type headers of the HTTP request. You can specify one of the following: or the default .

 <ul> <li> <p> <code>application/json</code> — standard JSON format </p> </li> <li> <p> <code>application/x-amzn-ion</code> — the Amazon Ion data format. For more information, see <a href="http://amzn.github.io/ion-docs/">Amazon Ion</a>. </p> </li> </ul>

$sel:profilingGroupName:PostAgentProfile', postAgentProfile_profilingGroupName - The name of the profiling group with the aggregated profile that receives the submitted profiling data.

Request Lenses

postAgentProfile_profileToken :: Lens' PostAgentProfile (Maybe Text)

Amazon CodeGuru Profiler uses this universally unique identifier (UUID) to prevent the accidental submission of duplicate profiling data if there are failures and retries.

postAgentProfile_agentProfile :: Lens' PostAgentProfile ByteString

The submitted profiling data.

postAgentProfile_contentType :: Lens' PostAgentProfile Text

The format of the submitted profiling data. The format maps to the Accept and Content-Type headers of the HTTP request. You can specify one of the following: or the default .

 <ul> <li> <p> <code>application/json</code> — standard JSON format </p> </li> <li> <p> <code>application/x-amzn-ion</code> — the Amazon Ion data format. For more information, see <a href="http://amzn.github.io/ion-docs/">Amazon Ion</a>. </p> </li> </ul>

postAgentProfile_profilingGroupName :: Lens' PostAgentProfile Text

The name of the profiling group with the aggregated profile that receives the submitted profiling data.

Destructuring the Response

data PostAgentProfileResponse

The structure representing the postAgentProfileResponse.

See: newPostAgentProfileResponse smart constructor.

Constructors

PostAgentProfileResponse'

Fields

httpStatus :: Int

The response's http status code.

newPostAgentProfileResponse

Arguments

:: Int
$sel:httpStatus:PostAgentProfileResponse'

-> PostAgentProfileResponse

Create a value of PostAgentProfileResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:httpStatus:PostAgentProfileResponse', postAgentProfileResponse_httpStatus - The response's http status code.

Response Lenses

postAgentProfileResponse_httpStatus :: Lens' PostAgentProfileResponse Int

The response's http status code.