Amazonka.CodeGuruProfiler.RemovePermission

Description

Removes permissions from a profiling group's resource-based policy that are provided using an action group. The one supported action group that can be removed is agentPermission which grants ConfigureAgent and PostAgent permissions. For more information, see Resource-based policies in CodeGuru Profiler in the Amazon CodeGuru Profiler User Guide, ConfigureAgent , and PostAgentProfile .

Creating a Request

data RemovePermission

 The structure representing the <code>removePermissionRequest</code>.</p>

See: newRemovePermission smart constructor.

Constructors

RemovePermission'

Fields

actionGroup :: ActionGroup

Specifies an action group that contains the permissions to remove from a profiling group's resource-based policy. One action group is supported, agentPermissions, which grants ConfigureAgent and PostAgentProfile permissions.

profilingGroupName :: Text

The name of the profiling group.

revisionId :: Text

A universally unique identifier (UUID) for the revision of the resource-based policy from which you want to remove permissions.

newRemovePermission

Arguments

:: ActionGroup
$sel:actionGroup:RemovePermission'

-> Text
$sel:profilingGroupName:RemovePermission'

-> Text
$sel:revisionId:RemovePermission'

-> RemovePermission

Create a value of RemovePermission with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:actionGroup:RemovePermission', removePermission_actionGroup - Specifies an action group that contains the permissions to remove from a profiling group's resource-based policy. One action group is supported, agentPermissions, which grants ConfigureAgent and PostAgentProfile permissions.

$sel:profilingGroupName:RemovePermission', removePermission_profilingGroupName - The name of the profiling group.

$sel:revisionId:RemovePermission', removePermission_revisionId - A universally unique identifier (UUID) for the revision of the resource-based policy from which you want to remove permissions.

Request Lenses

removePermission_actionGroup :: Lens' RemovePermission ActionGroup

Specifies an action group that contains the permissions to remove from a profiling group's resource-based policy. One action group is supported, agentPermissions, which grants ConfigureAgent and PostAgentProfile permissions.

removePermission_profilingGroupName :: Lens' RemovePermission Text

The name of the profiling group.

removePermission_revisionId :: Lens' RemovePermission Text

A universally unique identifier (UUID) for the revision of the resource-based policy from which you want to remove permissions.

Destructuring the Response

data RemovePermissionResponse

The structure representing the removePermissionResponse.

See: newRemovePermissionResponse smart constructor.

Constructors

RemovePermissionResponse'

Fields

httpStatus :: Int

The response's http status code.

policy :: Text

The JSON-formatted resource-based policy on the profiling group after the specified permissions were removed.

revisionId :: Text

A universally unique identifier (UUID) for the revision of the resource-based policy after the specified permissions were removed. The updated JSON-formatted policy is in the policy element of the response.

newRemovePermissionResponse

Arguments

:: Int
$sel:httpStatus:RemovePermissionResponse'

-> Text
$sel:policy:RemovePermissionResponse'

-> Text
$sel:revisionId:RemovePermission'

-> RemovePermissionResponse

Create a value of RemovePermissionResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:httpStatus:RemovePermissionResponse', removePermissionResponse_httpStatus - The response's http status code.

$sel:policy:RemovePermissionResponse', removePermissionResponse_policy - The JSON-formatted resource-based policy on the profiling group after the specified permissions were removed.

$sel:revisionId:RemovePermission', removePermissionResponse_revisionId - A universally unique identifier (UUID) for the revision of the resource-based policy after the specified permissions were removed. The updated JSON-formatted policy is in the policy element of the response.

Response Lenses

removePermissionResponse_httpStatus :: Lens' RemovePermissionResponse Int

The response's http status code.

removePermissionResponse_policy :: Lens' RemovePermissionResponse Text

The JSON-formatted resource-based policy on the profiling group after the specified permissions were removed.

removePermissionResponse_revisionId :: Lens' RemovePermissionResponse Text

A universally unique identifier (UUID) for the revision of the resource-based policy after the specified permissions were removed. The updated JSON-formatted policy is in the policy element of the response.