Amazonka.CodeGuruProfiler.Types

Description

Service Configuration

defaultService :: Service

API version 2019-07-18 of the Amazon CodeGuru Profiler SDK configuration.

Errors

_ValidationException :: AsError a => Getting (First ServiceError) a ServiceError

The parameter is not valid.

_ConflictException :: AsError a => Getting (First ServiceError) a ServiceError

The requested operation would cause a conflict with the current state of a service resource associated with the request. Resolve the conflict before retrying this request.

_ServiceQuotaExceededException :: AsError a => Getting (First ServiceError) a ServiceError

You have exceeded your service quota. To perform the requested action, remove some of the relevant resources, or use Service Quotas to request a service quota increase.

_ThrottlingException :: AsError a => Getting (First ServiceError) a ServiceError

The request was denied due to request throttling.

_InternalServerException :: AsError a => Getting (First ServiceError) a ServiceError

The server encountered an internal error and is unable to complete the request.

_ResourceNotFoundException :: AsError a => Getting (First ServiceError) a ServiceError

The resource specified in the request does not exist.

ActionGroup

newtype ActionGroup

Constructors

ActionGroup'

Fields

fromActionGroup :: Text

Bundled Patterns

pattern ActionGroup_AgentPermissions :: ActionGroup

AgentParameterField

newtype AgentParameterField

Constructors

AgentParameterField'

Fields

fromAgentParameterField :: Text

Bundled Patterns

pattern AgentParameterField_MaxStackDepth :: AgentParameterField

pattern AgentParameterField_MemoryUsageLimitPercent :: AgentParameterField

pattern AgentParameterField_MinimumTimeForReportingInMilliseconds :: AgentParameterField

pattern AgentParameterField_ReportingIntervalInMilliseconds :: AgentParameterField

pattern AgentParameterField_SamplingIntervalInMilliseconds :: AgentParameterField

AggregationPeriod

newtype AggregationPeriod

Constructors

AggregationPeriod'

Fields

fromAggregationPeriod :: Text

Bundled Patterns

pattern AggregationPeriod_P1D :: AggregationPeriod

pattern AggregationPeriod_PT1H :: AggregationPeriod

pattern AggregationPeriod_PT5M :: AggregationPeriod

ComputePlatform

newtype ComputePlatform

Constructors

ComputePlatform'

Fields

fromComputePlatform :: Text

Bundled Patterns

pattern ComputePlatform_AWSLambda :: ComputePlatform

pattern ComputePlatform_Default :: ComputePlatform

EventPublisher

newtype EventPublisher

Constructors

EventPublisher'

Fields

fromEventPublisher :: Text

Bundled Patterns

pattern EventPublisher_AnomalyDetection :: EventPublisher

FeedbackType

newtype FeedbackType

Constructors

FeedbackType'

Fields

fromFeedbackType :: Text

Bundled Patterns

pattern FeedbackType_Negative :: FeedbackType

pattern FeedbackType_Positive :: FeedbackType

MetadataField

newtype MetadataField

Constructors

MetadataField'

Fields

fromMetadataField :: Text

Bundled Patterns

pattern MetadataField_AgentId :: MetadataField

pattern MetadataField_AwsRequestId :: MetadataField

pattern MetadataField_ComputePlatform :: MetadataField

pattern MetadataField_ExecutionEnvironment :: MetadataField

pattern MetadataField_LambdaFunctionArn :: MetadataField

pattern MetadataField_LambdaMemoryLimitInMB :: MetadataField

pattern MetadataField_LambdaPreviousExecutionTimeInMilliseconds :: MetadataField

pattern MetadataField_LambdaRemainingTimeInMilliseconds :: MetadataField

pattern MetadataField_LambdaTimeGapBetweenInvokesInMilliseconds :: MetadataField

MetricType

newtype MetricType

Constructors

MetricType'

Fields

fromMetricType :: Text

Bundled Patterns

pattern MetricType_AggregatedRelativeTotalTime :: MetricType

OrderBy

newtype OrderBy

Constructors

OrderBy'

Fields

fromOrderBy :: Text

Bundled Patterns

pattern OrderBy_TimestampAscending :: OrderBy

pattern OrderBy_TimestampDescending :: OrderBy

AgentConfiguration

data AgentConfiguration

The response of ConfigureAgent that specifies if an agent profiles or not and for how long to return profiling data.

See: newAgentConfiguration smart constructor.

Constructors

AgentConfiguration'

Fields

agentParameters :: Maybe (HashMap AgentParameterField Text)

Parameters used by the profiler. The valid parameters are:

MaxStackDepth - The maximum depth of the stacks in the code that is represented in the profile. For example, if CodeGuru Profiler finds a method A, which calls method B, which calls method C, which calls method D, then the depth is 4. If the maxDepth is set to 2, then the profiler evaluates A and B.

MemoryUsageLimitPercent - The percentage of memory that is used by the profiler.

MinimumTimeForReportingInMilliseconds - The minimum time in milliseconds between sending reports.

ReportingIntervalInMilliseconds - The reporting interval in milliseconds used to report profiles.

SamplingIntervalInMilliseconds - The sampling interval in milliseconds that is used to profile samples.

periodInSeconds :: Int

How long a profiling agent should send profiling data using ConfigureAgent . For example, if this is set to 300, the profiling agent calls ConfigureAgent every 5 minutes to submit the profiled data collected during that period.

shouldProfile :: Bool

A Boolean that specifies whether the profiling agent collects profiling data or not. Set to true to enable profiling.

newAgentConfiguration

Arguments

:: Int
$sel:periodInSeconds:AgentConfiguration'

-> Bool
$sel:shouldProfile:AgentConfiguration'

-> AgentConfiguration

Create a value of AgentConfiguration with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:agentParameters:AgentConfiguration', agentConfiguration_agentParameters - Parameters used by the profiler. The valid parameters are:

MaxStackDepth - The maximum depth of the stacks in the code that is represented in the profile. For example, if CodeGuru Profiler finds a method A, which calls method B, which calls method C, which calls method D, then the depth is 4. If the maxDepth is set to 2, then the profiler evaluates A and B.

MemoryUsageLimitPercent - The percentage of memory that is used by the profiler.

MinimumTimeForReportingInMilliseconds - The minimum time in milliseconds between sending reports.

ReportingIntervalInMilliseconds - The reporting interval in milliseconds used to report profiles.

SamplingIntervalInMilliseconds - The sampling interval in milliseconds that is used to profile samples.

$sel:periodInSeconds:AgentConfiguration', agentConfiguration_periodInSeconds - How long a profiling agent should send profiling data using ConfigureAgent . For example, if this is set to 300, the profiling agent calls ConfigureAgent every 5 minutes to submit the profiled data collected during that period.

$sel:shouldProfile:AgentConfiguration', agentConfiguration_shouldProfile - A Boolean that specifies whether the profiling agent collects profiling data or not. Set to true to enable profiling.

agentConfiguration_agentParameters :: Lens' AgentConfiguration (Maybe (HashMap AgentParameterField Text))

Parameters used by the profiler. The valid parameters are:

MaxStackDepth - The maximum depth of the stacks in the code that is represented in the profile. For example, if CodeGuru Profiler finds a method A, which calls method B, which calls method C, which calls method D, then the depth is 4. If the maxDepth is set to 2, then the profiler evaluates A and B.

MemoryUsageLimitPercent - The percentage of memory that is used by the profiler.

MinimumTimeForReportingInMilliseconds - The minimum time in milliseconds between sending reports.

ReportingIntervalInMilliseconds - The reporting interval in milliseconds used to report profiles.

SamplingIntervalInMilliseconds - The sampling interval in milliseconds that is used to profile samples.

agentConfiguration_periodInSeconds :: Lens' AgentConfiguration Int

How long a profiling agent should send profiling data using ConfigureAgent . For example, if this is set to 300, the profiling agent calls ConfigureAgent every 5 minutes to submit the profiled data collected during that period.

agentConfiguration_shouldProfile :: Lens' AgentConfiguration Bool

A Boolean that specifies whether the profiling agent collects profiling data or not. Set to true to enable profiling.

AgentOrchestrationConfig

data AgentOrchestrationConfig

Specifies whether profiling is enabled or disabled for a profiling group. It is used by ConfigureAgent to enable or disable profiling for a profiling group.

See: newAgentOrchestrationConfig smart constructor.

Constructors

AgentOrchestrationConfig'

Fields

profilingEnabled :: Bool

A Boolean that specifies whether the profiling agent collects profiling data or not. Set to true to enable profiling.

newAgentOrchestrationConfig

Arguments

:: Bool
$sel:profilingEnabled:AgentOrchestrationConfig'

-> AgentOrchestrationConfig

Create a value of AgentOrchestrationConfig with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:profilingEnabled:AgentOrchestrationConfig', agentOrchestrationConfig_profilingEnabled - A Boolean that specifies whether the profiling agent collects profiling data or not. Set to true to enable profiling.

agentOrchestrationConfig_profilingEnabled :: Lens' AgentOrchestrationConfig Bool

A Boolean that specifies whether the profiling agent collects profiling data or not. Set to true to enable profiling.

AggregatedProfileTime

data AggregatedProfileTime

Specifies the aggregation period and aggregation start time for an aggregated profile. An aggregated profile is used to collect posted agent profiles during an aggregation period. There are three possible aggregation periods (1 day, 1 hour, or 5 minutes).

See: newAggregatedProfileTime smart constructor.

Constructors

AggregatedProfileTime'

Fields

period :: Maybe AggregationPeriod

The aggregation period. This indicates the period during which an aggregation profile collects posted agent profiles for a profiling group. Use one of three valid durations that are specified using the ISO 8601 format.

P1D — 1 day

PT1H — 1 hour

PT5M — 5 minutes

start :: Maybe POSIX

The time that aggregation of posted agent profiles for a profiling group starts. The aggregation profile contains profiles posted by the agent starting at this time for an aggregation period specified by the period property of the AggregatedProfileTime object.

Specify start using the ISO 8601 format. For example, 2020-06-01T13:15:02.001Z represents 1 millisecond past June 1, 2020 1:15:02 PM UTC.

newAggregatedProfileTime :: AggregatedProfileTime

Create a value of AggregatedProfileTime with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:period:AggregatedProfileTime', aggregatedProfileTime_period - The aggregation period. This indicates the period during which an aggregation profile collects posted agent profiles for a profiling group. Use one of three valid durations that are specified using the ISO 8601 format.

P1D — 1 day

PT1H — 1 hour

PT5M — 5 minutes

$sel:start:AggregatedProfileTime', aggregatedProfileTime_start - The time that aggregation of posted agent profiles for a profiling group starts. The aggregation profile contains profiles posted by the agent starting at this time for an aggregation period specified by the period property of the AggregatedProfileTime object.

Specify start using the ISO 8601 format. For example, 2020-06-01T13:15:02.001Z represents 1 millisecond past June 1, 2020 1:15:02 PM UTC.

aggregatedProfileTime_period :: Lens' AggregatedProfileTime (Maybe AggregationPeriod)

The aggregation period. This indicates the period during which an aggregation profile collects posted agent profiles for a profiling group. Use one of three valid durations that are specified using the ISO 8601 format.

P1D — 1 day

PT1H — 1 hour

PT5M — 5 minutes

aggregatedProfileTime_start :: Lens' AggregatedProfileTime (Maybe UTCTime)

The time that aggregation of posted agent profiles for a profiling group starts. The aggregation profile contains profiles posted by the agent starting at this time for an aggregation period specified by the period property of the AggregatedProfileTime object.

Specify start using the ISO 8601 format. For example, 2020-06-01T13:15:02.001Z represents 1 millisecond past June 1, 2020 1:15:02 PM UTC.

Anomaly

data Anomaly

Details about an anomaly in a specific metric of application profile. The anomaly is detected using analysis of the metric data over a period of time.

See: newAnomaly smart constructor.

Constructors

Anomaly'

Fields

instances :: [AnomalyInstance]

A list of the instances of the detected anomalies during the requested period.

metric :: Metric

Details about the metric that the analysis used when it detected the anomaly. The metric includes the name of the frame that was analyzed with the type and thread states used to derive the metric value for that frame.

reason :: Text

The reason for which metric was flagged as anomalous.

newAnomaly

Arguments

:: Metric
$sel:metric:Anomaly'

-> Text
$sel:reason:Anomaly'

-> Anomaly

Create a value of Anomaly with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:instances:Anomaly', anomaly_instances - A list of the instances of the detected anomalies during the requested period.

$sel:metric:Anomaly', anomaly_metric - Details about the metric that the analysis used when it detected the anomaly. The metric includes the name of the frame that was analyzed with the type and thread states used to derive the metric value for that frame.

$sel:reason:Anomaly', anomaly_reason - The reason for which metric was flagged as anomalous.

anomaly_instances :: Lens' Anomaly [AnomalyInstance]

A list of the instances of the detected anomalies during the requested period.

anomaly_metric :: Lens' Anomaly Metric

Details about the metric that the analysis used when it detected the anomaly. The metric includes the name of the frame that was analyzed with the type and thread states used to derive the metric value for that frame.

anomaly_reason :: Lens' Anomaly Text

The reason for which metric was flagged as anomalous.

AnomalyInstance

data AnomalyInstance

The specific duration in which the metric is flagged as anomalous.

See: newAnomalyInstance smart constructor.

Constructors

AnomalyInstance'

Fields

endTime :: Maybe POSIX

The end time of the period during which the metric is flagged as anomalous. This is specified using the ISO 8601 format. For example, 2020-06-01T13:15:02.001Z represents 1 millisecond past June 1, 2020 1:15:02 PM UTC.

userFeedback :: Maybe UserFeedback

Feedback type on a specific instance of anomaly submitted by the user.

id :: Text

The universally unique identifier (UUID) of an instance of an anomaly in a metric.

startTime :: POSIX

The start time of the period during which the metric is flagged as anomalous. This is specified using the ISO 8601 format. For example, 2020-06-01T13:15:02.001Z represents 1 millisecond past June 1, 2020 1:15:02 PM UTC.

newAnomalyInstance

Arguments

:: Text
$sel:id:AnomalyInstance'

-> UTCTime
$sel:startTime:AnomalyInstance'

-> AnomalyInstance

Create a value of AnomalyInstance with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:endTime:AnomalyInstance', anomalyInstance_endTime - The end time of the period during which the metric is flagged as anomalous. This is specified using the ISO 8601 format. For example, 2020-06-01T13:15:02.001Z represents 1 millisecond past June 1, 2020 1:15:02 PM UTC.

$sel:userFeedback:AnomalyInstance', anomalyInstance_userFeedback - Feedback type on a specific instance of anomaly submitted by the user.

$sel:id:AnomalyInstance', anomalyInstance_id - The universally unique identifier (UUID) of an instance of an anomaly in a metric.

$sel:startTime:AnomalyInstance', anomalyInstance_startTime - The start time of the period during which the metric is flagged as anomalous. This is specified using the ISO 8601 format. For example, 2020-06-01T13:15:02.001Z represents 1 millisecond past June 1, 2020 1:15:02 PM UTC.

anomalyInstance_endTime :: Lens' AnomalyInstance (Maybe UTCTime)

The end time of the period during which the metric is flagged as anomalous. This is specified using the ISO 8601 format. For example, 2020-06-01T13:15:02.001Z represents 1 millisecond past June 1, 2020 1:15:02 PM UTC.

anomalyInstance_userFeedback :: Lens' AnomalyInstance (Maybe UserFeedback)

Feedback type on a specific instance of anomaly submitted by the user.

anomalyInstance_id :: Lens' AnomalyInstance Text

The universally unique identifier (UUID) of an instance of an anomaly in a metric.

anomalyInstance_startTime :: Lens' AnomalyInstance UTCTime

The start time of the period during which the metric is flagged as anomalous. This is specified using the ISO 8601 format. For example, 2020-06-01T13:15:02.001Z represents 1 millisecond past June 1, 2020 1:15:02 PM UTC.

Channel

data Channel

Notification medium for users to get alerted for events that occur in application profile. We support SNS topic as a notification channel.

See: newChannel smart constructor.

Constructors

Channel'

Fields

id :: Maybe Text

Unique identifier for each Channel in the notification configuration of a Profiling Group. A random UUID for channelId is used when adding a channel to the notification configuration if not specified in the request.

eventPublishers :: NonEmpty EventPublisher

List of publishers for different type of events that may be detected in an application from the profile. Anomaly detection is the only event publisher in Profiler.

uri :: Text

Unique arn of the resource to be used for notifications. We support a valid SNS topic arn as a channel uri.

newChannel

Arguments

:: NonEmpty EventPublisher
$sel:eventPublishers:Channel'

-> Text
$sel:uri:Channel'

-> Channel

Create a value of Channel with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:id:Channel', channel_id - Unique identifier for each Channel in the notification configuration of a Profiling Group. A random UUID for channelId is used when adding a channel to the notification configuration if not specified in the request.

$sel:eventPublishers:Channel', channel_eventPublishers - List of publishers for different type of events that may be detected in an application from the profile. Anomaly detection is the only event publisher in Profiler.

$sel:uri:Channel', channel_uri - Unique arn of the resource to be used for notifications. We support a valid SNS topic arn as a channel uri.

channel_id :: Lens' Channel (Maybe Text)

Unique identifier for each Channel in the notification configuration of a Profiling Group. A random UUID for channelId is used when adding a channel to the notification configuration if not specified in the request.

channel_eventPublishers :: Lens' Channel (NonEmpty EventPublisher)

List of publishers for different type of events that may be detected in an application from the profile. Anomaly detection is the only event publisher in Profiler.

channel_uri :: Lens' Channel Text

Unique arn of the resource to be used for notifications. We support a valid SNS topic arn as a channel uri.

FindingsReportSummary

data FindingsReportSummary

Information about potential recommendations that might be created from the analysis of profiling data.

See: newFindingsReportSummary smart constructor.

Constructors

FindingsReportSummary'

Fields

profileStartTime :: Maybe POSIX

The start time of the profile the analysis data is about. This is specified using the ISO 8601 format. For example, 2020-06-01T13:15:02.001Z represents 1 millisecond past June 1, 2020 1:15:02 PM UTC.

profileEndTime :: Maybe POSIX

The end time of the period during which the metric is flagged as anomalous. This is specified using the ISO 8601 format. For example, 2020-06-01T13:15:02.001Z represents 1 millisecond past June 1, 2020 1:15:02 PM UTC.

id :: Maybe Text

The universally unique identifier (UUID) of the recommendation report.

profilingGroupName :: Maybe Text

The name of the profiling group that is associated with the analysis data.

totalNumberOfFindings :: Maybe Int

The total number of different recommendations that were found by the analysis.

newFindingsReportSummary :: FindingsReportSummary

Create a value of FindingsReportSummary with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:profileStartTime:FindingsReportSummary', findingsReportSummary_profileStartTime - The start time of the profile the analysis data is about. This is specified using the ISO 8601 format. For example, 2020-06-01T13:15:02.001Z represents 1 millisecond past June 1, 2020 1:15:02 PM UTC.

$sel:profileEndTime:FindingsReportSummary', findingsReportSummary_profileEndTime - The end time of the period during which the metric is flagged as anomalous. This is specified using the ISO 8601 format. For example, 2020-06-01T13:15:02.001Z represents 1 millisecond past June 1, 2020 1:15:02 PM UTC.

$sel:id:FindingsReportSummary', findingsReportSummary_id - The universally unique identifier (UUID) of the recommendation report.

$sel:profilingGroupName:FindingsReportSummary', findingsReportSummary_profilingGroupName - The name of the profiling group that is associated with the analysis data.

$sel:totalNumberOfFindings:FindingsReportSummary', findingsReportSummary_totalNumberOfFindings - The total number of different recommendations that were found by the analysis.

findingsReportSummary_profileStartTime :: Lens' FindingsReportSummary (Maybe UTCTime)

The start time of the profile the analysis data is about. This is specified using the ISO 8601 format. For example, 2020-06-01T13:15:02.001Z represents 1 millisecond past June 1, 2020 1:15:02 PM UTC.

findingsReportSummary_profileEndTime :: Lens' FindingsReportSummary (Maybe UTCTime)

The end time of the period during which the metric is flagged as anomalous. This is specified using the ISO 8601 format. For example, 2020-06-01T13:15:02.001Z represents 1 millisecond past June 1, 2020 1:15:02 PM UTC.

findingsReportSummary_id :: Lens' FindingsReportSummary (Maybe Text)

The universally unique identifier (UUID) of the recommendation report.

findingsReportSummary_profilingGroupName :: Lens' FindingsReportSummary (Maybe Text)

The name of the profiling group that is associated with the analysis data.

findingsReportSummary_totalNumberOfFindings :: Lens' FindingsReportSummary (Maybe Int)

The total number of different recommendations that were found by the analysis.

FrameMetric

data FrameMetric

The frame name, metric type, and thread states. These are used to derive the value of the metric for the frame.

See: newFrameMetric smart constructor.

Constructors

FrameMetric'

Fields

frameName :: Text

Name of the method common across the multiple occurrences of a frame in an application profile.

threadStates :: [Text]

List of application runtime thread states used to get the counts for a frame a derive a metric value.

type' :: MetricType

A type of aggregation that specifies how a metric for a frame is analyzed. The supported value AggregatedRelativeTotalTime is an aggregation of the metric value for one frame that is calculated across the occurrences of all frames in a profile.

newFrameMetric

Arguments

:: Text
$sel:frameName:FrameMetric'

-> MetricType
$sel:type':FrameMetric'

-> FrameMetric

Create a value of FrameMetric with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:frameName:FrameMetric', frameMetric_frameName - Name of the method common across the multiple occurrences of a frame in an application profile.

$sel:threadStates:FrameMetric', frameMetric_threadStates - List of application runtime thread states used to get the counts for a frame a derive a metric value.

$sel:type':FrameMetric', frameMetric_type - A type of aggregation that specifies how a metric for a frame is analyzed. The supported value AggregatedRelativeTotalTime is an aggregation of the metric value for one frame that is calculated across the occurrences of all frames in a profile.

frameMetric_frameName :: Lens' FrameMetric Text

Name of the method common across the multiple occurrences of a frame in an application profile.

frameMetric_threadStates :: Lens' FrameMetric [Text]

List of application runtime thread states used to get the counts for a frame a derive a metric value.

frameMetric_type :: Lens' FrameMetric MetricType

A type of aggregation that specifies how a metric for a frame is analyzed. The supported value AggregatedRelativeTotalTime is an aggregation of the metric value for one frame that is calculated across the occurrences of all frames in a profile.

FrameMetricDatum

data FrameMetricDatum

Information about a frame metric and its values.

See: newFrameMetricDatum smart constructor.

Constructors

FrameMetricDatum'

Fields

frameMetric :: FrameMetric

values :: [Double]

A list of values that are associated with a frame metric.

newFrameMetricDatum

Arguments

:: FrameMetric
$sel:frameMetric:FrameMetricDatum'

-> FrameMetricDatum

Create a value of FrameMetricDatum with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:frameMetric:FrameMetricDatum', frameMetricDatum_frameMetric - Undocumented member.

$sel:values:FrameMetricDatum', frameMetricDatum_values - A list of values that are associated with a frame metric.

frameMetricDatum_frameMetric :: Lens' FrameMetricDatum FrameMetric

Undocumented member.

frameMetricDatum_values :: Lens' FrameMetricDatum [Double]

A list of values that are associated with a frame metric.

Match

data Match

The part of a profile that contains a recommendation found during analysis.

See: newMatch smart constructor.

Constructors

Match'

Fields

thresholdBreachValue :: Maybe Double

The value in the profile data that exceeded the recommendation threshold.

frameAddress :: Maybe Text

The location in the profiling graph that contains a recommendation found during analysis.

targetFramesIndex :: Maybe Int

The target frame that triggered a match.

newMatch :: Match

Create a value of Match with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:thresholdBreachValue:Match', match_thresholdBreachValue - The value in the profile data that exceeded the recommendation threshold.

$sel:frameAddress:Match', match_frameAddress - The location in the profiling graph that contains a recommendation found during analysis.

$sel:targetFramesIndex:Match', match_targetFramesIndex - The target frame that triggered a match.

match_thresholdBreachValue :: Lens' Match (Maybe Double)

The value in the profile data that exceeded the recommendation threshold.

match_frameAddress :: Lens' Match (Maybe Text)

The location in the profiling graph that contains a recommendation found during analysis.

match_targetFramesIndex :: Lens' Match (Maybe Int)

The target frame that triggered a match.

Metric

data Metric

Details about the metric that the analysis used when it detected the anomaly. The metric what is analyzed to create recommendations. It includes the name of the frame that was analyzed and the type and thread states used to derive the metric value for that frame.

See: newMetric smart constructor.

Constructors

Metric'

Fields

frameName :: Text

The name of the method that appears as a frame in any stack in a profile.

threadStates :: [Text]

The list of application runtime thread states that is used to calculate the metric value for the frame.

type' :: MetricType

A type that specifies how a metric for a frame is analyzed. The supported value AggregatedRelativeTotalTime is an aggregation of the metric value for one frame that is calculated across the occurences of all frames in a profile.

newMetric

Arguments

:: Text
$sel:frameName:Metric'

-> MetricType
$sel:type':Metric'

-> Metric

Create a value of Metric with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:frameName:Metric', metric_frameName - The name of the method that appears as a frame in any stack in a profile.

$sel:threadStates:Metric', metric_threadStates - The list of application runtime thread states that is used to calculate the metric value for the frame.

$sel:type':Metric', metric_type - A type that specifies how a metric for a frame is analyzed. The supported value AggregatedRelativeTotalTime is an aggregation of the metric value for one frame that is calculated across the occurences of all frames in a profile.

metric_frameName :: Lens' Metric Text

The name of the method that appears as a frame in any stack in a profile.

metric_threadStates :: Lens' Metric [Text]

The list of application runtime thread states that is used to calculate the metric value for the frame.

metric_type :: Lens' Metric MetricType

A type that specifies how a metric for a frame is analyzed. The supported value AggregatedRelativeTotalTime is an aggregation of the metric value for one frame that is calculated across the occurences of all frames in a profile.

NotificationConfiguration

data NotificationConfiguration

The configuration for notifications stored for each profiling group. This includes up to to two channels and a list of event publishers associated with each channel.

See: newNotificationConfiguration smart constructor.

Constructors

NotificationConfiguration'

Fields

channels :: Maybe (NonEmpty Channel)

List of up to two channels to be used for sending notifications for events detected from the application profile.

newNotificationConfiguration :: NotificationConfiguration

Create a value of NotificationConfiguration with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:channels:NotificationConfiguration', notificationConfiguration_channels - List of up to two channels to be used for sending notifications for events detected from the application profile.

notificationConfiguration_channels :: Lens' NotificationConfiguration (Maybe (NonEmpty Channel))

List of up to two channels to be used for sending notifications for events detected from the application profile.

Pattern

data Pattern

A set of rules used to make a recommendation during an analysis.

See: newPattern smart constructor.

Constructors

Pattern'

Fields

thresholdPercent :: Maybe Double

The percentage of time an application spends in one method that triggers a recommendation. The percentage of time is the same as the percentage of the total gathered sample counts during analysis.

targetFrames :: Maybe [[Text]]

A list of frame names that were searched during the analysis that generated a recommendation.

countersToAggregate :: Maybe [Text]

A list of the different counters used to determine if there is a match.

name :: Maybe Text

The name for this pattern.

resolutionSteps :: Maybe Text

A string that contains the steps recommended to address the potential inefficiency.

id :: Maybe Text

The universally unique identifier (UUID) of this pattern.

description :: Maybe Text

The description of the recommendation. This explains a potential inefficiency in a profiled application.

newPattern :: Pattern

Create a value of Pattern with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:thresholdPercent:Pattern', pattern_thresholdPercent - The percentage of time an application spends in one method that triggers a recommendation. The percentage of time is the same as the percentage of the total gathered sample counts during analysis.

$sel:targetFrames:Pattern', pattern_targetFrames - A list of frame names that were searched during the analysis that generated a recommendation.

$sel:countersToAggregate:Pattern', pattern_countersToAggregate - A list of the different counters used to determine if there is a match.

$sel:name:Pattern', pattern_name - The name for this pattern.

$sel:resolutionSteps:Pattern', pattern_resolutionSteps - A string that contains the steps recommended to address the potential inefficiency.

$sel:id:Pattern', pattern_id - The universally unique identifier (UUID) of this pattern.

$sel:description:Pattern', pattern_description - The description of the recommendation. This explains a potential inefficiency in a profiled application.

pattern_thresholdPercent :: Lens' Pattern (Maybe Double)

The percentage of time an application spends in one method that triggers a recommendation. The percentage of time is the same as the percentage of the total gathered sample counts during analysis.

pattern_targetFrames :: Lens' Pattern (Maybe [[Text]])

A list of frame names that were searched during the analysis that generated a recommendation.

pattern_countersToAggregate :: Lens' Pattern (Maybe [Text])

A list of the different counters used to determine if there is a match.

pattern_name :: Lens' Pattern (Maybe Text)

The name for this pattern.

pattern_resolutionSteps :: Lens' Pattern (Maybe Text)

A string that contains the steps recommended to address the potential inefficiency.

pattern_id :: Lens' Pattern (Maybe Text)

The universally unique identifier (UUID) of this pattern.

pattern_description :: Lens' Pattern (Maybe Text)

The description of the recommendation. This explains a potential inefficiency in a profiled application.

ProfileTime

data ProfileTime

Contains the start time of a profile.

See: newProfileTime smart constructor.

Constructors

ProfileTime'

Fields

start :: Maybe POSIX

The start time of a profile. It is specified using the ISO 8601 format. For example, 2020-06-01T13:15:02.001Z represents 1 millisecond past June 1, 2020 1:15:02 PM UTC.

newProfileTime :: ProfileTime

Create a value of ProfileTime with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:start:ProfileTime', profileTime_start - The start time of a profile. It is specified using the ISO 8601 format. For example, 2020-06-01T13:15:02.001Z represents 1 millisecond past June 1, 2020 1:15:02 PM UTC.

profileTime_start :: Lens' ProfileTime (Maybe UTCTime)

The start time of a profile. It is specified using the ISO 8601 format. For example, 2020-06-01T13:15:02.001Z represents 1 millisecond past June 1, 2020 1:15:02 PM UTC.

ProfilingGroupDescription

data ProfilingGroupDescription

Contains information about a profiling group.

See: newProfilingGroupDescription smart constructor.

Constructors

ProfilingGroupDescription'

Fields

computePlatform :: Maybe ComputePlatform

The compute platform of the profiling group. If it is set to AWSLambda, then the profiled application runs on AWS Lambda. If it is set to Default, then the profiled application runs on a compute platform that is not AWS Lambda, such an Amazon EC2 instance, an on-premises server, or a different platform. The default is Default.

arn :: Maybe Text

The Amazon Resource Name (ARN) identifying the profiling group resource.

createdAt :: Maybe POSIX

The time when the profiling group was created. Specify using the ISO 8601 format. For example, 2020-06-01T13:15:02.001Z represents 1 millisecond past June 1, 2020 1:15:02 PM UTC.

name :: Maybe Text

The name of the profiling group.

profilingStatus :: Maybe ProfilingStatus

A ProfilingStatus object that includes information about the last time a profile agent pinged back, the last time a profile was received, and the aggregation period and start time for the most recent aggregated profile.

updatedAt :: Maybe POSIX

The date and time when the profiling group was last updated. Specify using the ISO 8601 format. For example, 2020-06-01T13:15:02.001Z represents 1 millisecond past June 1, 2020 1:15:02 PM UTC.

agentOrchestrationConfig :: Maybe AgentOrchestrationConfig

An AgentOrchestrationConfig object that indicates if the profiling group is enabled for profiled or not.

tags :: Maybe (HashMap Text Text)

A list of the tags that belong to this profiling group.

newProfilingGroupDescription :: ProfilingGroupDescription

Create a value of ProfilingGroupDescription with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:computePlatform:ProfilingGroupDescription', profilingGroupDescription_computePlatform - The compute platform of the profiling group. If it is set to AWSLambda, then the profiled application runs on AWS Lambda. If it is set to Default, then the profiled application runs on a compute platform that is not AWS Lambda, such an Amazon EC2 instance, an on-premises server, or a different platform. The default is Default.

$sel:arn:ProfilingGroupDescription', profilingGroupDescription_arn - The Amazon Resource Name (ARN) identifying the profiling group resource.

$sel:createdAt:ProfilingGroupDescription', profilingGroupDescription_createdAt - The time when the profiling group was created. Specify using the ISO 8601 format. For example, 2020-06-01T13:15:02.001Z represents 1 millisecond past June 1, 2020 1:15:02 PM UTC.

$sel:name:ProfilingGroupDescription', profilingGroupDescription_name - The name of the profiling group.

$sel:profilingStatus:ProfilingGroupDescription', profilingGroupDescription_profilingStatus - A ProfilingStatus object that includes information about the last time a profile agent pinged back, the last time a profile was received, and the aggregation period and start time for the most recent aggregated profile.

$sel:updatedAt:ProfilingGroupDescription', profilingGroupDescription_updatedAt - The date and time when the profiling group was last updated. Specify using the ISO 8601 format. For example, 2020-06-01T13:15:02.001Z represents 1 millisecond past June 1, 2020 1:15:02 PM UTC.

$sel:agentOrchestrationConfig:ProfilingGroupDescription', profilingGroupDescription_agentOrchestrationConfig - An AgentOrchestrationConfig object that indicates if the profiling group is enabled for profiled or not.

$sel:tags:ProfilingGroupDescription', profilingGroupDescription_tags - A list of the tags that belong to this profiling group.

profilingGroupDescription_computePlatform :: Lens' ProfilingGroupDescription (Maybe ComputePlatform)

The compute platform of the profiling group. If it is set to AWSLambda, then the profiled application runs on AWS Lambda. If it is set to Default, then the profiled application runs on a compute platform that is not AWS Lambda, such an Amazon EC2 instance, an on-premises server, or a different platform. The default is Default.

profilingGroupDescription_arn :: Lens' ProfilingGroupDescription (Maybe Text)

The Amazon Resource Name (ARN) identifying the profiling group resource.

profilingGroupDescription_createdAt :: Lens' ProfilingGroupDescription (Maybe UTCTime)

The time when the profiling group was created. Specify using the ISO 8601 format. For example, 2020-06-01T13:15:02.001Z represents 1 millisecond past June 1, 2020 1:15:02 PM UTC.

profilingGroupDescription_name :: Lens' ProfilingGroupDescription (Maybe Text)

The name of the profiling group.

profilingGroupDescription_profilingStatus :: Lens' ProfilingGroupDescription (Maybe ProfilingStatus)

A ProfilingStatus object that includes information about the last time a profile agent pinged back, the last time a profile was received, and the aggregation period and start time for the most recent aggregated profile.

profilingGroupDescription_updatedAt :: Lens' ProfilingGroupDescription (Maybe UTCTime)

The date and time when the profiling group was last updated. Specify using the ISO 8601 format. For example, 2020-06-01T13:15:02.001Z represents 1 millisecond past June 1, 2020 1:15:02 PM UTC.

profilingGroupDescription_agentOrchestrationConfig :: Lens' ProfilingGroupDescription (Maybe AgentOrchestrationConfig)

An AgentOrchestrationConfig object that indicates if the profiling group is enabled for profiled or not.

profilingGroupDescription_tags :: Lens' ProfilingGroupDescription (Maybe (HashMap Text Text))

A list of the tags that belong to this profiling group.

ProfilingStatus

data ProfilingStatus

Profiling status includes information about the last time a profile agent pinged back, the last time a profile was received, and the aggregation period and start time for the most recent aggregated profile.

See: newProfilingStatus smart constructor.

Constructors

ProfilingStatus'

Fields

latestAgentOrchestratedAt :: Maybe POSIX

The date and time when the profiling agent most recently pinged back. Specify using the ISO 8601 format. For example, 2020-06-01T13:15:02.001Z represents 1 millisecond past June 1, 2020 1:15:02 PM UTC.

latestAgentProfileReportedAt :: Maybe POSIX

The date and time when the most recent profile was received. Specify using the ISO 8601 format. For example, 2020-06-01T13:15:02.001Z represents 1 millisecond past June 1, 2020 1:15:02 PM UTC.

latestAggregatedProfile :: Maybe AggregatedProfileTime

An AggregatedProfileTime object that contains the aggregation period and start time for an aggregated profile.

newProfilingStatus :: ProfilingStatus

Create a value of ProfilingStatus with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:latestAgentOrchestratedAt:ProfilingStatus', profilingStatus_latestAgentOrchestratedAt - The date and time when the profiling agent most recently pinged back. Specify using the ISO 8601 format. For example, 2020-06-01T13:15:02.001Z represents 1 millisecond past June 1, 2020 1:15:02 PM UTC.

$sel:latestAgentProfileReportedAt:ProfilingStatus', profilingStatus_latestAgentProfileReportedAt - The date and time when the most recent profile was received. Specify using the ISO 8601 format. For example, 2020-06-01T13:15:02.001Z represents 1 millisecond past June 1, 2020 1:15:02 PM UTC.

$sel:latestAggregatedProfile:ProfilingStatus', profilingStatus_latestAggregatedProfile - An AggregatedProfileTime object that contains the aggregation period and start time for an aggregated profile.

profilingStatus_latestAgentOrchestratedAt :: Lens' ProfilingStatus (Maybe UTCTime)

The date and time when the profiling agent most recently pinged back. Specify using the ISO 8601 format. For example, 2020-06-01T13:15:02.001Z represents 1 millisecond past June 1, 2020 1:15:02 PM UTC.

profilingStatus_latestAgentProfileReportedAt :: Lens' ProfilingStatus (Maybe UTCTime)

The date and time when the most recent profile was received. Specify using the ISO 8601 format. For example, 2020-06-01T13:15:02.001Z represents 1 millisecond past June 1, 2020 1:15:02 PM UTC.

profilingStatus_latestAggregatedProfile :: Lens' ProfilingStatus (Maybe AggregatedProfileTime)

An AggregatedProfileTime object that contains the aggregation period and start time for an aggregated profile.

Recommendation

data Recommendation

A potential improvement that was found from analyzing the profiling data.

See: newRecommendation smart constructor.

Constructors

Recommendation'

Fields

allMatchesCount :: Int

How many different places in the profile graph triggered a match.

allMatchesSum :: Double

How much of the total sample count is potentially affected.

endTime :: POSIX

End time of the profile that was used by this analysis. This is specified using the ISO 8601 format. For example, 2020-06-01T13:15:02.001Z represents 1 millisecond past June 1, 2020 1:15:02 PM UTC.

pattern' :: Pattern

The pattern that analysis recognized in the profile to make this recommendation.

startTime :: POSIX

The start time of the profile that was used by this analysis. This is specified using the ISO 8601 format. For example, 2020-06-01T13:15:02.001Z represents 1 millisecond past June 1, 2020 1:15:02 PM UTC.

topMatches :: [Match]

List of the matches with most impact.

newRecommendation

Arguments

:: Int
$sel:allMatchesCount:Recommendation'

-> Double
$sel:allMatchesSum:Recommendation'

-> UTCTime
$sel:endTime:Recommendation'

-> Pattern
$sel:pattern':Recommendation'

-> UTCTime
$sel:startTime:Recommendation'

-> Recommendation

Create a value of Recommendation with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:allMatchesCount:Recommendation', recommendation_allMatchesCount - How many different places in the profile graph triggered a match.

$sel:allMatchesSum:Recommendation', recommendation_allMatchesSum - How much of the total sample count is potentially affected.

$sel:endTime:Recommendation', recommendation_endTime - End time of the profile that was used by this analysis. This is specified using the ISO 8601 format. For example, 2020-06-01T13:15:02.001Z represents 1 millisecond past June 1, 2020 1:15:02 PM UTC.

$sel:pattern':Recommendation', recommendation_pattern - The pattern that analysis recognized in the profile to make this recommendation.

$sel:startTime:Recommendation', recommendation_startTime - The start time of the profile that was used by this analysis. This is specified using the ISO 8601 format. For example, 2020-06-01T13:15:02.001Z represents 1 millisecond past June 1, 2020 1:15:02 PM UTC.

$sel:topMatches:Recommendation', recommendation_topMatches - List of the matches with most impact.

recommendation_allMatchesCount :: Lens' Recommendation Int

How many different places in the profile graph triggered a match.

recommendation_allMatchesSum :: Lens' Recommendation Double

How much of the total sample count is potentially affected.

recommendation_endTime :: Lens' Recommendation UTCTime

End time of the profile that was used by this analysis. This is specified using the ISO 8601 format. For example, 2020-06-01T13:15:02.001Z represents 1 millisecond past June 1, 2020 1:15:02 PM UTC.

recommendation_pattern :: Lens' Recommendation Pattern

The pattern that analysis recognized in the profile to make this recommendation.

recommendation_startTime :: Lens' Recommendation UTCTime

The start time of the profile that was used by this analysis. This is specified using the ISO 8601 format. For example, 2020-06-01T13:15:02.001Z represents 1 millisecond past June 1, 2020 1:15:02 PM UTC.

recommendation_topMatches :: Lens' Recommendation [Match]

List of the matches with most impact.

TimestampStructure

data TimestampStructure

A data type that contains a Timestamp object. This is specified using the ISO 8601 format. For example, 2020-06-01T13:15:02.001Z represents 1 millisecond past June 1, 2020 1:15:02 PM UTC.

See: newTimestampStructure smart constructor.

Constructors

TimestampStructure'

Fields

value :: POSIX

A Timestamp. This is specified using the ISO 8601 format. For example, 2020-06-01T13:15:02.001Z represents 1 millisecond past June 1, 2020 1:15:02 PM UTC.

newTimestampStructure

Arguments

:: UTCTime
$sel:value:TimestampStructure'

-> TimestampStructure

Create a value of TimestampStructure with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:value:TimestampStructure', timestampStructure_value - A Timestamp. This is specified using the ISO 8601 format. For example, 2020-06-01T13:15:02.001Z represents 1 millisecond past June 1, 2020 1:15:02 PM UTC.

timestampStructure_value :: Lens' TimestampStructure UTCTime

A Timestamp. This is specified using the ISO 8601 format. For example, 2020-06-01T13:15:02.001Z represents 1 millisecond past June 1, 2020 1:15:02 PM UTC.

UserFeedback

data UserFeedback

Feedback that can be submitted for each instance of an anomaly by the user. Feedback is be used for improvements in generating recommendations for the application.

See: newUserFeedback smart constructor.

Constructors

UserFeedback'

Fields

type' :: FeedbackType

Optional Positive or Negative feedback submitted by the user about whether the recommendation is useful or not.

newUserFeedback

Arguments

:: FeedbackType
$sel:type':UserFeedback'

-> UserFeedback

Create a value of UserFeedback with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:type':UserFeedback', userFeedback_type - Optional Positive or Negative feedback submitted by the user about whether the recommendation is useful or not.

userFeedback_type :: Lens' UserFeedback FeedbackType

Optional Positive or Negative feedback submitted by the user about whether the recommendation is useful or not.