Amazonka.CodeGuruProfiler.Types.AgentConfiguration

Description

Documentation

data AgentConfiguration

The response of ConfigureAgent that specifies if an agent profiles or not and for how long to return profiling data.

See: newAgentConfiguration smart constructor.

Constructors

AgentConfiguration'

Fields

agentParameters :: Maybe (HashMap AgentParameterField Text)

Parameters used by the profiler. The valid parameters are:

MaxStackDepth - The maximum depth of the stacks in the code that is represented in the profile. For example, if CodeGuru Profiler finds a method A, which calls method B, which calls method C, which calls method D, then the depth is 4. If the maxDepth is set to 2, then the profiler evaluates A and B.

MemoryUsageLimitPercent - The percentage of memory that is used by the profiler.

MinimumTimeForReportingInMilliseconds - The minimum time in milliseconds between sending reports.

ReportingIntervalInMilliseconds - The reporting interval in milliseconds used to report profiles.

SamplingIntervalInMilliseconds - The sampling interval in milliseconds that is used to profile samples.

periodInSeconds :: Int

How long a profiling agent should send profiling data using ConfigureAgent . For example, if this is set to 300, the profiling agent calls ConfigureAgent every 5 minutes to submit the profiled data collected during that period.

shouldProfile :: Bool

A Boolean that specifies whether the profiling agent collects profiling data or not. Set to true to enable profiling.

newAgentConfiguration

Arguments

:: Int
$sel:periodInSeconds:AgentConfiguration'

-> Bool
$sel:shouldProfile:AgentConfiguration'

-> AgentConfiguration

Create a value of AgentConfiguration with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:agentParameters:AgentConfiguration', agentConfiguration_agentParameters - Parameters used by the profiler. The valid parameters are:

MaxStackDepth - The maximum depth of the stacks in the code that is represented in the profile. For example, if CodeGuru Profiler finds a method A, which calls method B, which calls method C, which calls method D, then the depth is 4. If the maxDepth is set to 2, then the profiler evaluates A and B.

MemoryUsageLimitPercent - The percentage of memory that is used by the profiler.

MinimumTimeForReportingInMilliseconds - The minimum time in milliseconds between sending reports.

ReportingIntervalInMilliseconds - The reporting interval in milliseconds used to report profiles.

SamplingIntervalInMilliseconds - The sampling interval in milliseconds that is used to profile samples.

$sel:periodInSeconds:AgentConfiguration', agentConfiguration_periodInSeconds - How long a profiling agent should send profiling data using ConfigureAgent . For example, if this is set to 300, the profiling agent calls ConfigureAgent every 5 minutes to submit the profiled data collected during that period.

$sel:shouldProfile:AgentConfiguration', agentConfiguration_shouldProfile - A Boolean that specifies whether the profiling agent collects profiling data or not. Set to true to enable profiling.

agentConfiguration_agentParameters :: Lens' AgentConfiguration (Maybe (HashMap AgentParameterField Text))

Parameters used by the profiler. The valid parameters are:

MaxStackDepth - The maximum depth of the stacks in the code that is represented in the profile. For example, if CodeGuru Profiler finds a method A, which calls method B, which calls method C, which calls method D, then the depth is 4. If the maxDepth is set to 2, then the profiler evaluates A and B.

MemoryUsageLimitPercent - The percentage of memory that is used by the profiler.

MinimumTimeForReportingInMilliseconds - The minimum time in milliseconds between sending reports.

ReportingIntervalInMilliseconds - The reporting interval in milliseconds used to report profiles.

SamplingIntervalInMilliseconds - The sampling interval in milliseconds that is used to profile samples.

agentConfiguration_periodInSeconds :: Lens' AgentConfiguration Int

How long a profiling agent should send profiling data using ConfigureAgent . For example, if this is set to 300, the profiling agent calls ConfigureAgent every 5 minutes to submit the profiled data collected during that period.

agentConfiguration_shouldProfile :: Lens' AgentConfiguration Bool

A Boolean that specifies whether the profiling agent collects profiling data or not. Set to true to enable profiling.