Amazonka.CodeGuruProfiler.Types.AgentOrchestrationConfig

Description

Documentation

data AgentOrchestrationConfig

Specifies whether profiling is enabled or disabled for a profiling group. It is used by ConfigureAgent to enable or disable profiling for a profiling group.

See: newAgentOrchestrationConfig smart constructor.

Constructors

AgentOrchestrationConfig'

Fields

profilingEnabled :: Bool

A Boolean that specifies whether the profiling agent collects profiling data or not. Set to true to enable profiling.

newAgentOrchestrationConfig

Arguments

:: Bool
$sel:profilingEnabled:AgentOrchestrationConfig'

-> AgentOrchestrationConfig

Create a value of AgentOrchestrationConfig with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:profilingEnabled:AgentOrchestrationConfig', agentOrchestrationConfig_profilingEnabled - A Boolean that specifies whether the profiling agent collects profiling data or not. Set to true to enable profiling.

agentOrchestrationConfig_profilingEnabled :: Lens' AgentOrchestrationConfig Bool

A Boolean that specifies whether the profiling agent collects profiling data or not. Set to true to enable profiling.