Amazonka.CodeGuruProfiler.Types.AggregatedProfileTime

Description

Documentation

data AggregatedProfileTime

Specifies the aggregation period and aggregation start time for an aggregated profile. An aggregated profile is used to collect posted agent profiles during an aggregation period. There are three possible aggregation periods (1 day, 1 hour, or 5 minutes).

See: newAggregatedProfileTime smart constructor.

Constructors

AggregatedProfileTime'

Fields

period :: Maybe AggregationPeriod

The aggregation period. This indicates the period during which an aggregation profile collects posted agent profiles for a profiling group. Use one of three valid durations that are specified using the ISO 8601 format.

P1D — 1 day

PT1H — 1 hour

PT5M — 5 minutes

start :: Maybe POSIX

The time that aggregation of posted agent profiles for a profiling group starts. The aggregation profile contains profiles posted by the agent starting at this time for an aggregation period specified by the period property of the AggregatedProfileTime object.

Specify start using the ISO 8601 format. For example, 2020-06-01T13:15:02.001Z represents 1 millisecond past June 1, 2020 1:15:02 PM UTC.

newAggregatedProfileTime :: AggregatedProfileTime

Create a value of AggregatedProfileTime with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:period:AggregatedProfileTime', aggregatedProfileTime_period - The aggregation period. This indicates the period during which an aggregation profile collects posted agent profiles for a profiling group. Use one of three valid durations that are specified using the ISO 8601 format.

P1D — 1 day

PT1H — 1 hour

PT5M — 5 minutes

$sel:start:AggregatedProfileTime', aggregatedProfileTime_start - The time that aggregation of posted agent profiles for a profiling group starts. The aggregation profile contains profiles posted by the agent starting at this time for an aggregation period specified by the period property of the AggregatedProfileTime object.

Specify start using the ISO 8601 format. For example, 2020-06-01T13:15:02.001Z represents 1 millisecond past June 1, 2020 1:15:02 PM UTC.

aggregatedProfileTime_period :: Lens' AggregatedProfileTime (Maybe AggregationPeriod)

The aggregation period. This indicates the period during which an aggregation profile collects posted agent profiles for a profiling group. Use one of three valid durations that are specified using the ISO 8601 format.

P1D — 1 day

PT1H — 1 hour

PT5M — 5 minutes

aggregatedProfileTime_start :: Lens' AggregatedProfileTime (Maybe UTCTime)

The time that aggregation of posted agent profiles for a profiling group starts. The aggregation profile contains profiles posted by the agent starting at this time for an aggregation period specified by the period property of the AggregatedProfileTime object.

Specify start using the ISO 8601 format. For example, 2020-06-01T13:15:02.001Z represents 1 millisecond past June 1, 2020 1:15:02 PM UTC.