Amazonka.CodeGuruProfiler.Types.FrameMetric

Description

Documentation

data FrameMetric

The frame name, metric type, and thread states. These are used to derive the value of the metric for the frame.

See: newFrameMetric smart constructor.

Constructors

FrameMetric'

Fields

frameName :: Text

Name of the method common across the multiple occurrences of a frame in an application profile.

threadStates :: [Text]

List of application runtime thread states used to get the counts for a frame a derive a metric value.

type' :: MetricType

A type of aggregation that specifies how a metric for a frame is analyzed. The supported value AggregatedRelativeTotalTime is an aggregation of the metric value for one frame that is calculated across the occurrences of all frames in a profile.

newFrameMetric

Arguments

:: Text
$sel:frameName:FrameMetric'

-> MetricType
$sel:type':FrameMetric'

-> FrameMetric

Create a value of FrameMetric with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:frameName:FrameMetric', frameMetric_frameName - Name of the method common across the multiple occurrences of a frame in an application profile.

$sel:threadStates:FrameMetric', frameMetric_threadStates - List of application runtime thread states used to get the counts for a frame a derive a metric value.

$sel:type':FrameMetric', frameMetric_type - A type of aggregation that specifies how a metric for a frame is analyzed. The supported value AggregatedRelativeTotalTime is an aggregation of the metric value for one frame that is calculated across the occurrences of all frames in a profile.

frameMetric_frameName :: Lens' FrameMetric Text

Name of the method common across the multiple occurrences of a frame in an application profile.

frameMetric_threadStates :: Lens' FrameMetric [Text]

List of application runtime thread states used to get the counts for a frame a derive a metric value.

frameMetric_type :: Lens' FrameMetric MetricType

A type of aggregation that specifies how a metric for a frame is analyzed. The supported value AggregatedRelativeTotalTime is an aggregation of the metric value for one frame that is calculated across the occurrences of all frames in a profile.