Amazonka.CodeGuruProfiler.Types.ProfilingGroupDescription

Description

Documentation

data ProfilingGroupDescription

Contains information about a profiling group.

See: newProfilingGroupDescription smart constructor.

Constructors

ProfilingGroupDescription'

Fields

computePlatform :: Maybe ComputePlatform

The compute platform of the profiling group. If it is set to AWSLambda, then the profiled application runs on AWS Lambda. If it is set to Default, then the profiled application runs on a compute platform that is not AWS Lambda, such an Amazon EC2 instance, an on-premises server, or a different platform. The default is Default.

arn :: Maybe Text

The Amazon Resource Name (ARN) identifying the profiling group resource.

createdAt :: Maybe POSIX

The time when the profiling group was created. Specify using the ISO 8601 format. For example, 2020-06-01T13:15:02.001Z represents 1 millisecond past June 1, 2020 1:15:02 PM UTC.

name :: Maybe Text

The name of the profiling group.

profilingStatus :: Maybe ProfilingStatus

A ProfilingStatus object that includes information about the last time a profile agent pinged back, the last time a profile was received, and the aggregation period and start time for the most recent aggregated profile.

updatedAt :: Maybe POSIX

The date and time when the profiling group was last updated. Specify using the ISO 8601 format. For example, 2020-06-01T13:15:02.001Z represents 1 millisecond past June 1, 2020 1:15:02 PM UTC.

agentOrchestrationConfig :: Maybe AgentOrchestrationConfig

An AgentOrchestrationConfig object that indicates if the profiling group is enabled for profiled or not.

tags :: Maybe (HashMap Text Text)

A list of the tags that belong to this profiling group.

newProfilingGroupDescription :: ProfilingGroupDescription

Create a value of ProfilingGroupDescription with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:computePlatform:ProfilingGroupDescription', profilingGroupDescription_computePlatform - The compute platform of the profiling group. If it is set to AWSLambda, then the profiled application runs on AWS Lambda. If it is set to Default, then the profiled application runs on a compute platform that is not AWS Lambda, such an Amazon EC2 instance, an on-premises server, or a different platform. The default is Default.

$sel:arn:ProfilingGroupDescription', profilingGroupDescription_arn - The Amazon Resource Name (ARN) identifying the profiling group resource.

$sel:createdAt:ProfilingGroupDescription', profilingGroupDescription_createdAt - The time when the profiling group was created. Specify using the ISO 8601 format. For example, 2020-06-01T13:15:02.001Z represents 1 millisecond past June 1, 2020 1:15:02 PM UTC.

$sel:name:ProfilingGroupDescription', profilingGroupDescription_name - The name of the profiling group.

$sel:profilingStatus:ProfilingGroupDescription', profilingGroupDescription_profilingStatus - A ProfilingStatus object that includes information about the last time a profile agent pinged back, the last time a profile was received, and the aggregation period and start time for the most recent aggregated profile.

$sel:updatedAt:ProfilingGroupDescription', profilingGroupDescription_updatedAt - The date and time when the profiling group was last updated. Specify using the ISO 8601 format. For example, 2020-06-01T13:15:02.001Z represents 1 millisecond past June 1, 2020 1:15:02 PM UTC.

$sel:agentOrchestrationConfig:ProfilingGroupDescription', profilingGroupDescription_agentOrchestrationConfig - An AgentOrchestrationConfig object that indicates if the profiling group is enabled for profiled or not.

$sel:tags:ProfilingGroupDescription', profilingGroupDescription_tags - A list of the tags that belong to this profiling group.

profilingGroupDescription_computePlatform :: Lens' ProfilingGroupDescription (Maybe ComputePlatform)

The compute platform of the profiling group. If it is set to AWSLambda, then the profiled application runs on AWS Lambda. If it is set to Default, then the profiled application runs on a compute platform that is not AWS Lambda, such an Amazon EC2 instance, an on-premises server, or a different platform. The default is Default.

profilingGroupDescription_arn :: Lens' ProfilingGroupDescription (Maybe Text)

The Amazon Resource Name (ARN) identifying the profiling group resource.

profilingGroupDescription_createdAt :: Lens' ProfilingGroupDescription (Maybe UTCTime)

The time when the profiling group was created. Specify using the ISO 8601 format. For example, 2020-06-01T13:15:02.001Z represents 1 millisecond past June 1, 2020 1:15:02 PM UTC.

profilingGroupDescription_name :: Lens' ProfilingGroupDescription (Maybe Text)

The name of the profiling group.

profilingGroupDescription_profilingStatus :: Lens' ProfilingGroupDescription (Maybe ProfilingStatus)

A ProfilingStatus object that includes information about the last time a profile agent pinged back, the last time a profile was received, and the aggregation period and start time for the most recent aggregated profile.

profilingGroupDescription_updatedAt :: Lens' ProfilingGroupDescription (Maybe UTCTime)

The date and time when the profiling group was last updated. Specify using the ISO 8601 format. For example, 2020-06-01T13:15:02.001Z represents 1 millisecond past June 1, 2020 1:15:02 PM UTC.

profilingGroupDescription_agentOrchestrationConfig :: Lens' ProfilingGroupDescription (Maybe AgentOrchestrationConfig)

An AgentOrchestrationConfig object that indicates if the profiling group is enabled for profiled or not.

profilingGroupDescription_tags :: Lens' ProfilingGroupDescription (Maybe (HashMap Text Text))

A list of the tags that belong to this profiling group.