Amazonka.CodeGuruReviewer.AssociateRepository

Description

Use to associate an Amazon Web Services CodeCommit repository or a repostory managed by Amazon Web Services CodeStar Connections with Amazon CodeGuru Reviewer. When you associate a repository, CodeGuru Reviewer reviews source code changes in the repository's pull requests and provides automatic recommendations. You can view recommendations using the CodeGuru Reviewer console. For more information, see Recommendations in Amazon CodeGuru Reviewer in the Amazon CodeGuru Reviewer User Guide.

If you associate a CodeCommit or S3 repository, it must be in the same Amazon Web Services Region and Amazon Web Services account where its CodeGuru Reviewer code reviews are configured.

Bitbucket and GitHub Enterprise Server repositories are managed by Amazon Web Services CodeStar Connections to connect to CodeGuru Reviewer. For more information, see Associate a repository in the Amazon CodeGuru Reviewer User Guide.

You cannot use the CodeGuru Reviewer SDK or the Amazon Web Services CLI to associate a GitHub repository with Amazon CodeGuru Reviewer. To associate a GitHub repository, use the console. For more information, see Getting started with CodeGuru Reviewer in the CodeGuru Reviewer User Guide.

Creating a Request

data AssociateRepository

See: newAssociateRepository smart constructor.

Constructors

AssociateRepository'

Fields

kmsKeyDetails :: Maybe KMSKeyDetails

A KMSKeyDetails object that contains:

The encryption option for this repository association. It is either owned by Amazon Web Services Key Management Service (KMS) (AWS_OWNED_CMK) or customer managed (CUSTOMER_MANAGED_CMK).

The ID of the Amazon Web Services KMS key that is associated with this respository association.

clientRequestToken :: Maybe Text

Amazon CodeGuru Reviewer uses this value to prevent the accidental creation of duplicate repository associations if there are failures and retries.

tags :: Maybe (HashMap Text Text)

An array of key-value pairs used to tag an associated repository. A tag is a custom attribute label with two parts:

A tag key (for example, CostCenter, Environment, Project, or Secret). Tag keys are case sensitive.

An optional field known as a tag value (for example, 111122223333, Production, or a team name). Omitting the tag value is the same as using an empty string. Like tag keys, tag values are case sensitive.

repository :: Repository

The repository to associate.

newAssociateRepository

Arguments

:: Repository
$sel:repository:AssociateRepository'

-> AssociateRepository

Create a value of AssociateRepository with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:kmsKeyDetails:AssociateRepository', associateRepository_kmsKeyDetails - A KMSKeyDetails object that contains:

The encryption option for this repository association. It is either owned by Amazon Web Services Key Management Service (KMS) (AWS_OWNED_CMK) or customer managed (CUSTOMER_MANAGED_CMK).

The ID of the Amazon Web Services KMS key that is associated with this respository association.

$sel:clientRequestToken:AssociateRepository', associateRepository_clientRequestToken - Amazon CodeGuru Reviewer uses this value to prevent the accidental creation of duplicate repository associations if there are failures and retries.

$sel:tags:AssociateRepository', associateRepository_tags - An array of key-value pairs used to tag an associated repository. A tag is a custom attribute label with two parts:

A tag key (for example, CostCenter, Environment, Project, or Secret). Tag keys are case sensitive.

An optional field known as a tag value (for example, 111122223333, Production, or a team name). Omitting the tag value is the same as using an empty string. Like tag keys, tag values are case sensitive.

$sel:repository:AssociateRepository', associateRepository_repository - The repository to associate.

Request Lenses

associateRepository_kmsKeyDetails :: Lens' AssociateRepository (Maybe KMSKeyDetails)

A KMSKeyDetails object that contains:

The encryption option for this repository association. It is either owned by Amazon Web Services Key Management Service (KMS) (AWS_OWNED_CMK) or customer managed (CUSTOMER_MANAGED_CMK).

The ID of the Amazon Web Services KMS key that is associated with this respository association.

associateRepository_clientRequestToken :: Lens' AssociateRepository (Maybe Text)

Amazon CodeGuru Reviewer uses this value to prevent the accidental creation of duplicate repository associations if there are failures and retries.

associateRepository_tags :: Lens' AssociateRepository (Maybe (HashMap Text Text))

An array of key-value pairs used to tag an associated repository. A tag is a custom attribute label with two parts:

A tag key (for example, CostCenter, Environment, Project, or Secret). Tag keys are case sensitive.

An optional field known as a tag value (for example, 111122223333, Production, or a team name). Omitting the tag value is the same as using an empty string. Like tag keys, tag values are case sensitive.

associateRepository_repository :: Lens' AssociateRepository Repository

The repository to associate.

Destructuring the Response

data AssociateRepositoryResponse

See: newAssociateRepositoryResponse smart constructor.

Constructors

AssociateRepositoryResponse'

Fields

repositoryAssociation :: Maybe RepositoryAssociation

Information about the repository association.

tags :: Maybe (HashMap Text Text)

An array of key-value pairs used to tag an associated repository. A tag is a custom attribute label with two parts:

A tag key (for example, CostCenter, Environment, Project, or Secret). Tag keys are case sensitive.

An optional field known as a tag value (for example, 111122223333, Production, or a team name). Omitting the tag value is the same as using an empty string. Like tag keys, tag values are case sensitive.

httpStatus :: Int

The response's http status code.

newAssociateRepositoryResponse

Arguments

:: Int
$sel:httpStatus:AssociateRepositoryResponse'

-> AssociateRepositoryResponse

Create a value of AssociateRepositoryResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:repositoryAssociation:AssociateRepositoryResponse', associateRepositoryResponse_repositoryAssociation - Information about the repository association.

$sel:tags:AssociateRepository', associateRepositoryResponse_tags - An array of key-value pairs used to tag an associated repository. A tag is a custom attribute label with two parts:

A tag key (for example, CostCenter, Environment, Project, or Secret). Tag keys are case sensitive.

An optional field known as a tag value (for example, 111122223333, Production, or a team name). Omitting the tag value is the same as using an empty string. Like tag keys, tag values are case sensitive.

$sel:httpStatus:AssociateRepositoryResponse', associateRepositoryResponse_httpStatus - The response's http status code.

Response Lenses

associateRepositoryResponse_repositoryAssociation :: Lens' AssociateRepositoryResponse (Maybe RepositoryAssociation)

Information about the repository association.

associateRepositoryResponse_tags :: Lens' AssociateRepositoryResponse (Maybe (HashMap Text Text))

An array of key-value pairs used to tag an associated repository. A tag is a custom attribute label with two parts:

A tag key (for example, CostCenter, Environment, Project, or Secret). Tag keys are case sensitive.

An optional field known as a tag value (for example, 111122223333, Production, or a team name). Omitting the tag value is the same as using an empty string. Like tag keys, tag values are case sensitive.

associateRepositoryResponse_httpStatus :: Lens' AssociateRepositoryResponse Int

The response's http status code.