Amazonka.CodeGuruReviewer.ListRepositoryAssociations

Description

Returns a list of RepositoryAssociationSummary objects that contain summary information about a repository association. You can filter the returned list by ProviderType , Name , State , and Owner .

This operation returns paginated results.

Creating a Request

data ListRepositoryAssociations

See: newListRepositoryAssociations smart constructor.

Constructors

ListRepositoryAssociations'

Fields

states :: Maybe (NonEmpty RepositoryAssociationState)

List of repository association states to use as a filter.

The valid repository association states are:

Associated: The repository association is complete.

Associating: CodeGuru Reviewer is:

Setting up pull request notifications. This is required for pull requests to trigger a CodeGuru Reviewer review.

If your repository ProviderType is GitHub, GitHub Enterprise Server, or Bitbucket, CodeGuru Reviewer creates webhooks in your repository to trigger CodeGuru Reviewer reviews. If you delete these webhooks, reviews of code in your repository cannot be triggered.

Setting up source code access. This is required for CodeGuru Reviewer to securely clone code in your repository.

Failed: The repository failed to associate or disassociate.

Disassociating: CodeGuru Reviewer is removing the repository's pull request notifications and source code access.

Disassociated: CodeGuru Reviewer successfully disassociated the repository. You can create a new association with this repository if you want to review source code in it later. You can control access to code reviews created in an associated repository with tags after it has been disassociated. For more information, see Using tags to control access to associated repositories in the Amazon CodeGuru Reviewer User Guide.

owners :: Maybe (NonEmpty Text)

List of owners to use as a filter. For Amazon Web Services CodeCommit, it is the name of the CodeCommit account that was used to associate the repository. For other repository source providers, such as Bitbucket and GitHub Enterprise Server, this is name of the account that was used to associate the repository.

providerTypes :: Maybe (NonEmpty ProviderType)

List of provider types to use as a filter.

nextToken :: Maybe Text

The nextToken value returned from a previous paginated ListRepositoryAssociations request where maxResults was used and the results exceeded the value of that parameter. Pagination continues from the end of the previous results that returned the nextToken value.

Treat this token as an opaque identifier that is only used to retrieve the next items in a list and not for other programmatic purposes.

names :: Maybe (NonEmpty Text)

List of repository names to use as a filter.

maxResults :: Maybe Natural

The maximum number of repository association results returned by ListRepositoryAssociations in paginated output. When this parameter is used, ListRepositoryAssociations only returns maxResults results in a single page with a nextToken response element. The remaining results of the initial request can be seen by sending another ListRepositoryAssociations request with the returned nextToken value. This value can be between 1 and 100. If this parameter is not used, ListRepositoryAssociations returns up to 100 results and a nextToken value if applicable.

newListRepositoryAssociations :: ListRepositoryAssociations

Create a value of ListRepositoryAssociations with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:states:ListRepositoryAssociations', listRepositoryAssociations_states - List of repository association states to use as a filter.

The valid repository association states are:

Associated: The repository association is complete.

Associating: CodeGuru Reviewer is:

Setting up pull request notifications. This is required for pull requests to trigger a CodeGuru Reviewer review.

If your repository ProviderType is GitHub, GitHub Enterprise Server, or Bitbucket, CodeGuru Reviewer creates webhooks in your repository to trigger CodeGuru Reviewer reviews. If you delete these webhooks, reviews of code in your repository cannot be triggered.

Setting up source code access. This is required for CodeGuru Reviewer to securely clone code in your repository.

Failed: The repository failed to associate or disassociate.

Disassociating: CodeGuru Reviewer is removing the repository's pull request notifications and source code access.

Disassociated: CodeGuru Reviewer successfully disassociated the repository. You can create a new association with this repository if you want to review source code in it later. You can control access to code reviews created in an associated repository with tags after it has been disassociated. For more information, see Using tags to control access to associated repositories in the Amazon CodeGuru Reviewer User Guide.

$sel:owners:ListRepositoryAssociations', listRepositoryAssociations_owners - List of owners to use as a filter. For Amazon Web Services CodeCommit, it is the name of the CodeCommit account that was used to associate the repository. For other repository source providers, such as Bitbucket and GitHub Enterprise Server, this is name of the account that was used to associate the repository.

$sel:providerTypes:ListRepositoryAssociations', listRepositoryAssociations_providerTypes - List of provider types to use as a filter.

$sel:nextToken:ListRepositoryAssociations', listRepositoryAssociations_nextToken - The nextToken value returned from a previous paginated ListRepositoryAssociations request where maxResults was used and the results exceeded the value of that parameter. Pagination continues from the end of the previous results that returned the nextToken value.

Treat this token as an opaque identifier that is only used to retrieve the next items in a list and not for other programmatic purposes.

$sel:names:ListRepositoryAssociations', listRepositoryAssociations_names - List of repository names to use as a filter.

$sel:maxResults:ListRepositoryAssociations', listRepositoryAssociations_maxResults - The maximum number of repository association results returned by ListRepositoryAssociations in paginated output. When this parameter is used, ListRepositoryAssociations only returns maxResults results in a single page with a nextToken response element. The remaining results of the initial request can be seen by sending another ListRepositoryAssociations request with the returned nextToken value. This value can be between 1 and 100. If this parameter is not used, ListRepositoryAssociations returns up to 100 results and a nextToken value if applicable.

Request Lenses

listRepositoryAssociations_states :: Lens' ListRepositoryAssociations (Maybe (NonEmpty RepositoryAssociationState))

List of repository association states to use as a filter.

The valid repository association states are:

Associated: The repository association is complete.

Associating: CodeGuru Reviewer is:

Setting up pull request notifications. This is required for pull requests to trigger a CodeGuru Reviewer review.

If your repository ProviderType is GitHub, GitHub Enterprise Server, or Bitbucket, CodeGuru Reviewer creates webhooks in your repository to trigger CodeGuru Reviewer reviews. If you delete these webhooks, reviews of code in your repository cannot be triggered.

Setting up source code access. This is required for CodeGuru Reviewer to securely clone code in your repository.

Failed: The repository failed to associate or disassociate.

Disassociating: CodeGuru Reviewer is removing the repository's pull request notifications and source code access.

Disassociated: CodeGuru Reviewer successfully disassociated the repository. You can create a new association with this repository if you want to review source code in it later. You can control access to code reviews created in an associated repository with tags after it has been disassociated. For more information, see Using tags to control access to associated repositories in the Amazon CodeGuru Reviewer User Guide.

listRepositoryAssociations_owners :: Lens' ListRepositoryAssociations (Maybe (NonEmpty Text))

List of owners to use as a filter. For Amazon Web Services CodeCommit, it is the name of the CodeCommit account that was used to associate the repository. For other repository source providers, such as Bitbucket and GitHub Enterprise Server, this is name of the account that was used to associate the repository.

listRepositoryAssociations_providerTypes :: Lens' ListRepositoryAssociations (Maybe (NonEmpty ProviderType))

List of provider types to use as a filter.

listRepositoryAssociations_nextToken :: Lens' ListRepositoryAssociations (Maybe Text)

The nextToken value returned from a previous paginated ListRepositoryAssociations request where maxResults was used and the results exceeded the value of that parameter. Pagination continues from the end of the previous results that returned the nextToken value.

Treat this token as an opaque identifier that is only used to retrieve the next items in a list and not for other programmatic purposes.

listRepositoryAssociations_names :: Lens' ListRepositoryAssociations (Maybe (NonEmpty Text))

List of repository names to use as a filter.

listRepositoryAssociations_maxResults :: Lens' ListRepositoryAssociations (Maybe Natural)

The maximum number of repository association results returned by ListRepositoryAssociations in paginated output. When this parameter is used, ListRepositoryAssociations only returns maxResults results in a single page with a nextToken response element. The remaining results of the initial request can be seen by sending another ListRepositoryAssociations request with the returned nextToken value. This value can be between 1 and 100. If this parameter is not used, ListRepositoryAssociations returns up to 100 results and a nextToken value if applicable.

Destructuring the Response

data ListRepositoryAssociationsResponse

See: newListRepositoryAssociationsResponse smart constructor.

Constructors

ListRepositoryAssociationsResponse'

Fields

nextToken :: Maybe Text

The nextToken value to include in a future ListRecommendations request. When the results of a ListRecommendations request exceed maxResults, this value can be used to retrieve the next page of results. This value is null when there are no more results to return.

repositoryAssociationSummaries :: Maybe [RepositoryAssociationSummary]

A list of repository associations that meet the criteria of the request.

httpStatus :: Int

The response's http status code.

newListRepositoryAssociationsResponse

Arguments

:: Int
$sel:httpStatus:ListRepositoryAssociationsResponse'

-> ListRepositoryAssociationsResponse

Create a value of ListRepositoryAssociationsResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:nextToken:ListRepositoryAssociations', listRepositoryAssociationsResponse_nextToken - The nextToken value to include in a future ListRecommendations request. When the results of a ListRecommendations request exceed maxResults, this value can be used to retrieve the next page of results. This value is null when there are no more results to return.

$sel:repositoryAssociationSummaries:ListRepositoryAssociationsResponse', listRepositoryAssociationsResponse_repositoryAssociationSummaries - A list of repository associations that meet the criteria of the request.

$sel:httpStatus:ListRepositoryAssociationsResponse', listRepositoryAssociationsResponse_httpStatus - The response's http status code.

Response Lenses

listRepositoryAssociationsResponse_nextToken :: Lens' ListRepositoryAssociationsResponse (Maybe Text)

The nextToken value to include in a future ListRecommendations request. When the results of a ListRecommendations request exceed maxResults, this value can be used to retrieve the next page of results. This value is null when there are no more results to return.

listRepositoryAssociationsResponse_repositoryAssociationSummaries :: Lens' ListRepositoryAssociationsResponse (Maybe [RepositoryAssociationSummary])

A list of repository associations that meet the criteria of the request.

listRepositoryAssociationsResponse_httpStatus :: Lens' ListRepositoryAssociationsResponse Int

The response's http status code.