Amazonka.CodeGuruReviewer.Types.CodeReviewSummary

Description

Documentation

data CodeReviewSummary

Information about the summary of the code review.

See: newCodeReviewSummary smart constructor.

Constructors

CodeReviewSummary'

Fields

state :: Maybe JobState

The state of the code review.

The valid code review states are:

Completed: The code review is complete.

Pending: The code review started and has not completed or failed.

Failed: The code review failed.

Deleting: The code review is being deleted.

pullRequestId :: Maybe Text

The pull request ID for the code review.

providerType :: Maybe ProviderType

The provider type of the repository association.

owner :: Maybe Text

The owner of the repository. For an Amazon Web Services CodeCommit repository, this is the Amazon Web Services account ID of the account that owns the repository. For a GitHub, GitHub Enterprise Server, or Bitbucket repository, this is the username for the account that owns the repository. For an S3 repository, it can be the username or Amazon Web Services account ID.

name :: Maybe Text

The name of the code review.

codeReviewArn :: Maybe Text

The Amazon Resource Name (ARN) of the CodeReview object.

repositoryName :: Maybe Text

The name of the repository.

type' :: Maybe Type

The type of the code review.

sourceCodeType :: Maybe SourceCodeType

metricsSummary :: Maybe MetricsSummary

The statistics from the code review.

createdTimeStamp :: Maybe POSIX

The time, in milliseconds since the epoch, when the code review was created.

lastUpdatedTimeStamp :: Maybe POSIX

The time, in milliseconds since the epoch, when the code review was last updated.

newCodeReviewSummary :: CodeReviewSummary

Create a value of CodeReviewSummary with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:state:CodeReviewSummary', codeReviewSummary_state - The state of the code review.

The valid code review states are:

Completed: The code review is complete.

Pending: The code review started and has not completed or failed.

Failed: The code review failed.

Deleting: The code review is being deleted.

$sel:pullRequestId:CodeReviewSummary', codeReviewSummary_pullRequestId - The pull request ID for the code review.

$sel:providerType:CodeReviewSummary', codeReviewSummary_providerType - The provider type of the repository association.

$sel:owner:CodeReviewSummary', codeReviewSummary_owner - The owner of the repository. For an Amazon Web Services CodeCommit repository, this is the Amazon Web Services account ID of the account that owns the repository. For a GitHub, GitHub Enterprise Server, or Bitbucket repository, this is the username for the account that owns the repository. For an S3 repository, it can be the username or Amazon Web Services account ID.

$sel:name:CodeReviewSummary', codeReviewSummary_name - The name of the code review.

$sel:codeReviewArn:CodeReviewSummary', codeReviewSummary_codeReviewArn - The Amazon Resource Name (ARN) of the CodeReview object.

$sel:repositoryName:CodeReviewSummary', codeReviewSummary_repositoryName - The name of the repository.

$sel:type':CodeReviewSummary', codeReviewSummary_type - The type of the code review.

$sel:sourceCodeType:CodeReviewSummary', codeReviewSummary_sourceCodeType - Undocumented member.

$sel:metricsSummary:CodeReviewSummary', codeReviewSummary_metricsSummary - The statistics from the code review.

$sel:createdTimeStamp:CodeReviewSummary', codeReviewSummary_createdTimeStamp - The time, in milliseconds since the epoch, when the code review was created.

$sel:lastUpdatedTimeStamp:CodeReviewSummary', codeReviewSummary_lastUpdatedTimeStamp - The time, in milliseconds since the epoch, when the code review was last updated.

codeReviewSummary_state :: Lens' CodeReviewSummary (Maybe JobState)

The state of the code review.

The valid code review states are:

Completed: The code review is complete.

Pending: The code review started and has not completed or failed.

Failed: The code review failed.

Deleting: The code review is being deleted.

codeReviewSummary_pullRequestId :: Lens' CodeReviewSummary (Maybe Text)

The pull request ID for the code review.

codeReviewSummary_providerType :: Lens' CodeReviewSummary (Maybe ProviderType)

The provider type of the repository association.

codeReviewSummary_owner :: Lens' CodeReviewSummary (Maybe Text)

The owner of the repository. For an Amazon Web Services CodeCommit repository, this is the Amazon Web Services account ID of the account that owns the repository. For a GitHub, GitHub Enterprise Server, or Bitbucket repository, this is the username for the account that owns the repository. For an S3 repository, it can be the username or Amazon Web Services account ID.

codeReviewSummary_name :: Lens' CodeReviewSummary (Maybe Text)

The name of the code review.

codeReviewSummary_codeReviewArn :: Lens' CodeReviewSummary (Maybe Text)

The Amazon Resource Name (ARN) of the CodeReview object.

codeReviewSummary_repositoryName :: Lens' CodeReviewSummary (Maybe Text)

The name of the repository.

codeReviewSummary_type :: Lens' CodeReviewSummary (Maybe Type)

The type of the code review.

codeReviewSummary_sourceCodeType :: Lens' CodeReviewSummary (Maybe SourceCodeType)

Undocumented member.

codeReviewSummary_metricsSummary :: Lens' CodeReviewSummary (Maybe MetricsSummary)

The statistics from the code review.

codeReviewSummary_createdTimeStamp :: Lens' CodeReviewSummary (Maybe UTCTime)

The time, in milliseconds since the epoch, when the code review was created.

codeReviewSummary_lastUpdatedTimeStamp :: Lens' CodeReviewSummary (Maybe UTCTime)

The time, in milliseconds since the epoch, when the code review was last updated.