Amazonka.CodeGuruReviewer.Types.CommitDiffSourceCodeType

Description

Documentation

data CommitDiffSourceCodeType

A type of SourceCodeType that specifies the commit diff for a pull request on an associated repository. The SourceCommit and DestinationCommit fields are required to do a pull request code review.

See: newCommitDiffSourceCodeType smart constructor.

Constructors

CommitDiffSourceCodeType'

Fields

sourceCommit :: Maybe Text

The SHA of the source commit used to generate a commit diff. This field is required for a pull request code review.

mergeBaseCommit :: Maybe Text

The SHA of the merge base of a commit.

destinationCommit :: Maybe Text

The SHA of the destination commit used to generate a commit diff. This field is required for a pull request code review.

newCommitDiffSourceCodeType :: CommitDiffSourceCodeType

Create a value of CommitDiffSourceCodeType with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:sourceCommit:CommitDiffSourceCodeType', commitDiffSourceCodeType_sourceCommit - The SHA of the source commit used to generate a commit diff. This field is required for a pull request code review.

$sel:mergeBaseCommit:CommitDiffSourceCodeType', commitDiffSourceCodeType_mergeBaseCommit - The SHA of the merge base of a commit.

$sel:destinationCommit:CommitDiffSourceCodeType', commitDiffSourceCodeType_destinationCommit - The SHA of the destination commit used to generate a commit diff. This field is required for a pull request code review.

commitDiffSourceCodeType_sourceCommit :: Lens' CommitDiffSourceCodeType (Maybe Text)

The SHA of the source commit used to generate a commit diff. This field is required for a pull request code review.

commitDiffSourceCodeType_mergeBaseCommit :: Lens' CommitDiffSourceCodeType (Maybe Text)

The SHA of the merge base of a commit.

commitDiffSourceCodeType_destinationCommit :: Lens' CommitDiffSourceCodeType (Maybe Text)

The SHA of the destination commit used to generate a commit diff. This field is required for a pull request code review.