Amazonka.CodeGuruReviewer.Types.MetricsSummary

Description

Documentation

data MetricsSummary

Information about metrics summaries.

See: newMetricsSummary smart constructor.

Constructors

MetricsSummary'

Fields

findingsCount :: Maybe Integer

Total number of recommendations found in the code review.

meteredLinesOfCodeCount :: Maybe Integer

Lines of code metered in the code review. For the initial code review pull request and all subsequent revisions, this includes all lines of code in the files added to the pull request. In subsequent revisions, for files that already existed in the pull request, this includes only the changed lines of code. In both cases, this does not include non-code lines such as comments and import statements. For example, if you submit a pull request containing 5 files, each with 500 lines of code, and in a subsequent revision you added a new file with 200 lines of code, and also modified a total of 25 lines across the initial 5 files, MeteredLinesOfCodeCount includes the first 5 files (5 * 500 = 2,500 lines), the new file (200 lines) and the 25 changed lines of code for a total of 2,725 lines of code.

newMetricsSummary :: MetricsSummary

Create a value of MetricsSummary with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:findingsCount:MetricsSummary', metricsSummary_findingsCount - Total number of recommendations found in the code review.

$sel:meteredLinesOfCodeCount:MetricsSummary', metricsSummary_meteredLinesOfCodeCount - Lines of code metered in the code review. For the initial code review pull request and all subsequent revisions, this includes all lines of code in the files added to the pull request. In subsequent revisions, for files that already existed in the pull request, this includes only the changed lines of code. In both cases, this does not include non-code lines such as comments and import statements. For example, if you submit a pull request containing 5 files, each with 500 lines of code, and in a subsequent revision you added a new file with 200 lines of code, and also modified a total of 25 lines across the initial 5 files, MeteredLinesOfCodeCount includes the first 5 files (5 * 500 = 2,500 lines), the new file (200 lines) and the 25 changed lines of code for a total of 2,725 lines of code.

metricsSummary_findingsCount :: Lens' MetricsSummary (Maybe Integer)

Total number of recommendations found in the code review.

metricsSummary_meteredLinesOfCodeCount :: Lens' MetricsSummary (Maybe Integer)

Lines of code metered in the code review. For the initial code review pull request and all subsequent revisions, this includes all lines of code in the files added to the pull request. In subsequent revisions, for files that already existed in the pull request, this includes only the changed lines of code. In both cases, this does not include non-code lines such as comments and import statements. For example, if you submit a pull request containing 5 files, each with 500 lines of code, and in a subsequent revision you added a new file with 200 lines of code, and also modified a total of 25 lines across the initial 5 files, MeteredLinesOfCodeCount includes the first 5 files (5 * 500 = 2,500 lines), the new file (200 lines) and the 25 changed lines of code for a total of 2,725 lines of code.