Amazonka.CodeGuruReviewer.Types.RecommendationFeedback

Description

Documentation

data RecommendationFeedback

Information about the recommendation feedback.

See: newRecommendationFeedback smart constructor.

Constructors

RecommendationFeedback'

Fields

recommendationId :: Maybe Text

The recommendation ID that can be used to track the provided recommendations. Later on it can be used to collect the feedback.

userId :: Maybe Text

The ID of the user that made the API call.

The UserId is an IAM principal that can be specified as an Amazon Web Services account ID or an Amazon Resource Name (ARN). For more information, see Specifying a Principal in the Amazon Web Services Identity and Access Management User Guide.

reactions :: Maybe [Reaction]

List for storing reactions. Reactions are utf-8 text code for emojis. You can send an empty list to clear off all your feedback.

codeReviewArn :: Maybe Text

The Amazon Resource Name (ARN) of the CodeReview object.

createdTimeStamp :: Maybe POSIX

The time at which the feedback was created.

lastUpdatedTimeStamp :: Maybe POSIX

The time at which the feedback was last updated.

newRecommendationFeedback :: RecommendationFeedback

Create a value of RecommendationFeedback with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:recommendationId:RecommendationFeedback', recommendationFeedback_recommendationId - The recommendation ID that can be used to track the provided recommendations. Later on it can be used to collect the feedback.

$sel:userId:RecommendationFeedback', recommendationFeedback_userId - The ID of the user that made the API call.

The UserId is an IAM principal that can be specified as an Amazon Web Services account ID or an Amazon Resource Name (ARN). For more information, see Specifying a Principal in the Amazon Web Services Identity and Access Management User Guide.

$sel:reactions:RecommendationFeedback', recommendationFeedback_reactions - List for storing reactions. Reactions are utf-8 text code for emojis. You can send an empty list to clear off all your feedback.

$sel:codeReviewArn:RecommendationFeedback', recommendationFeedback_codeReviewArn - The Amazon Resource Name (ARN) of the CodeReview object.

$sel:createdTimeStamp:RecommendationFeedback', recommendationFeedback_createdTimeStamp - The time at which the feedback was created.

$sel:lastUpdatedTimeStamp:RecommendationFeedback', recommendationFeedback_lastUpdatedTimeStamp - The time at which the feedback was last updated.

recommendationFeedback_recommendationId :: Lens' RecommendationFeedback (Maybe Text)

The recommendation ID that can be used to track the provided recommendations. Later on it can be used to collect the feedback.

recommendationFeedback_userId :: Lens' RecommendationFeedback (Maybe Text)

The ID of the user that made the API call.

The UserId is an IAM principal that can be specified as an Amazon Web Services account ID or an Amazon Resource Name (ARN). For more information, see Specifying a Principal in the Amazon Web Services Identity and Access Management User Guide.

recommendationFeedback_reactions :: Lens' RecommendationFeedback (Maybe [Reaction])

List for storing reactions. Reactions are utf-8 text code for emojis. You can send an empty list to clear off all your feedback.

recommendationFeedback_codeReviewArn :: Lens' RecommendationFeedback (Maybe Text)

The Amazon Resource Name (ARN) of the CodeReview object.

recommendationFeedback_createdTimeStamp :: Lens' RecommendationFeedback (Maybe UTCTime)

The time at which the feedback was created.

recommendationFeedback_lastUpdatedTimeStamp :: Lens' RecommendationFeedback (Maybe UTCTime)

The time at which the feedback was last updated.