Amazonka.CodeGuruReviewer.Types.RepositoryAssociation

Description

Documentation

data RepositoryAssociation

Information about a repository association. The DescribeRepositoryAssociation operation returns a RepositoryAssociation object.

See: newRepositoryAssociation smart constructor.

Constructors

RepositoryAssociation'

Fields

associationArn :: Maybe Text

The Amazon Resource Name (ARN) identifying the repository association.

associationId :: Maybe Text

The ID of the repository association.

state :: Maybe RepositoryAssociationState

The state of the repository association.

The valid repository association states are:

Associated: The repository association is complete.

Associating: CodeGuru Reviewer is:

Setting up pull request notifications. This is required for pull requests to trigger a CodeGuru Reviewer review.

If your repository ProviderType is GitHub, GitHub Enterprise Server, or Bitbucket, CodeGuru Reviewer creates webhooks in your repository to trigger CodeGuru Reviewer reviews. If you delete these webhooks, reviews of code in your repository cannot be triggered.

Setting up source code access. This is required for CodeGuru Reviewer to securely clone code in your repository.

Failed: The repository failed to associate or disassociate.

Disassociating: CodeGuru Reviewer is removing the repository's pull request notifications and source code access.

Disassociated: CodeGuru Reviewer successfully disassociated the repository. You can create a new association with this repository if you want to review source code in it later. You can control access to code reviews created in an associated repository with tags after it has been disassociated. For more information, see Using tags to control access to associated repositories in the Amazon CodeGuru Reviewer User Guide.

s3RepositoryDetails :: Maybe S3RepositoryDetails

providerType :: Maybe ProviderType

The provider type of the repository association.

owner :: Maybe Text

The owner of the repository. For an Amazon Web Services CodeCommit repository, this is the Amazon Web Services account ID of the account that owns the repository. For a GitHub, GitHub Enterprise Server, or Bitbucket repository, this is the username for the account that owns the repository. For an S3 repository, it can be the username or Amazon Web Services account ID.

name :: Maybe Text

The name of the repository.

kmsKeyDetails :: Maybe KMSKeyDetails

A KMSKeyDetails object that contains:

The encryption option for this repository association. It is either owned by Amazon Web Services Key Management Service (KMS) (AWS_OWNED_CMK) or customer managed (CUSTOMER_MANAGED_CMK).

The ID of the Amazon Web Services KMS key that is associated with this respository association.

connectionArn :: Maybe Text

The Amazon Resource Name (ARN) of an Amazon Web Services CodeStar Connections connection. Its format is arn:aws:codestar-connections:region-id:aws-account_id:connection/connection-id. For more information, see Connection in the Amazon Web Services CodeStar Connections API Reference.

stateReason :: Maybe Text

A description of why the repository association is in the current state.

createdTimeStamp :: Maybe POSIX

The time, in milliseconds since the epoch, when the repository association was created.

lastUpdatedTimeStamp :: Maybe POSIX

The time, in milliseconds since the epoch, when the repository association was last updated.

newRepositoryAssociation :: RepositoryAssociation

Create a value of RepositoryAssociation with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:associationArn:RepositoryAssociation', repositoryAssociation_associationArn - The Amazon Resource Name (ARN) identifying the repository association.

$sel:associationId:RepositoryAssociation', repositoryAssociation_associationId - The ID of the repository association.

$sel:state:RepositoryAssociation', repositoryAssociation_state - The state of the repository association.

The valid repository association states are:

Associated: The repository association is complete.

Associating: CodeGuru Reviewer is:

Setting up pull request notifications. This is required for pull requests to trigger a CodeGuru Reviewer review.

If your repository ProviderType is GitHub, GitHub Enterprise Server, or Bitbucket, CodeGuru Reviewer creates webhooks in your repository to trigger CodeGuru Reviewer reviews. If you delete these webhooks, reviews of code in your repository cannot be triggered.

Setting up source code access. This is required for CodeGuru Reviewer to securely clone code in your repository.

Failed: The repository failed to associate or disassociate.

Disassociating: CodeGuru Reviewer is removing the repository's pull request notifications and source code access.

Disassociated: CodeGuru Reviewer successfully disassociated the repository. You can create a new association with this repository if you want to review source code in it later. You can control access to code reviews created in an associated repository with tags after it has been disassociated. For more information, see Using tags to control access to associated repositories in the Amazon CodeGuru Reviewer User Guide.

$sel:s3RepositoryDetails:RepositoryAssociation', repositoryAssociation_s3RepositoryDetails - Undocumented member.

$sel:providerType:RepositoryAssociation', repositoryAssociation_providerType - The provider type of the repository association.

$sel:owner:RepositoryAssociation', repositoryAssociation_owner - The owner of the repository. For an Amazon Web Services CodeCommit repository, this is the Amazon Web Services account ID of the account that owns the repository. For a GitHub, GitHub Enterprise Server, or Bitbucket repository, this is the username for the account that owns the repository. For an S3 repository, it can be the username or Amazon Web Services account ID.

$sel:name:RepositoryAssociation', repositoryAssociation_name - The name of the repository.

$sel:kmsKeyDetails:RepositoryAssociation', repositoryAssociation_kmsKeyDetails - A KMSKeyDetails object that contains:

The encryption option for this repository association. It is either owned by Amazon Web Services Key Management Service (KMS) (AWS_OWNED_CMK) or customer managed (CUSTOMER_MANAGED_CMK).

The ID of the Amazon Web Services KMS key that is associated with this respository association.

$sel:connectionArn:RepositoryAssociation', repositoryAssociation_connectionArn - The Amazon Resource Name (ARN) of an Amazon Web Services CodeStar Connections connection. Its format is arn:aws:codestar-connections:region-id:aws-account_id:connection/connection-id. For more information, see Connection in the Amazon Web Services CodeStar Connections API Reference.

$sel:stateReason:RepositoryAssociation', repositoryAssociation_stateReason - A description of why the repository association is in the current state.

$sel:createdTimeStamp:RepositoryAssociation', repositoryAssociation_createdTimeStamp - The time, in milliseconds since the epoch, when the repository association was created.

$sel:lastUpdatedTimeStamp:RepositoryAssociation', repositoryAssociation_lastUpdatedTimeStamp - The time, in milliseconds since the epoch, when the repository association was last updated.

repositoryAssociation_associationArn :: Lens' RepositoryAssociation (Maybe Text)

The Amazon Resource Name (ARN) identifying the repository association.

repositoryAssociation_associationId :: Lens' RepositoryAssociation (Maybe Text)

The ID of the repository association.

repositoryAssociation_state :: Lens' RepositoryAssociation (Maybe RepositoryAssociationState)

The state of the repository association.

The valid repository association states are:

Associated: The repository association is complete.

Associating: CodeGuru Reviewer is:

Setting up pull request notifications. This is required for pull requests to trigger a CodeGuru Reviewer review.

If your repository ProviderType is GitHub, GitHub Enterprise Server, or Bitbucket, CodeGuru Reviewer creates webhooks in your repository to trigger CodeGuru Reviewer reviews. If you delete these webhooks, reviews of code in your repository cannot be triggered.

Setting up source code access. This is required for CodeGuru Reviewer to securely clone code in your repository.

Failed: The repository failed to associate or disassociate.

Disassociating: CodeGuru Reviewer is removing the repository's pull request notifications and source code access.

Disassociated: CodeGuru Reviewer successfully disassociated the repository. You can create a new association with this repository if you want to review source code in it later. You can control access to code reviews created in an associated repository with tags after it has been disassociated. For more information, see Using tags to control access to associated repositories in the Amazon CodeGuru Reviewer User Guide.

repositoryAssociation_s3RepositoryDetails :: Lens' RepositoryAssociation (Maybe S3RepositoryDetails)

Undocumented member.

repositoryAssociation_providerType :: Lens' RepositoryAssociation (Maybe ProviderType)

The provider type of the repository association.

repositoryAssociation_owner :: Lens' RepositoryAssociation (Maybe Text)

The owner of the repository. For an Amazon Web Services CodeCommit repository, this is the Amazon Web Services account ID of the account that owns the repository. For a GitHub, GitHub Enterprise Server, or Bitbucket repository, this is the username for the account that owns the repository. For an S3 repository, it can be the username or Amazon Web Services account ID.

repositoryAssociation_name :: Lens' RepositoryAssociation (Maybe Text)

The name of the repository.

repositoryAssociation_kmsKeyDetails :: Lens' RepositoryAssociation (Maybe KMSKeyDetails)

A KMSKeyDetails object that contains:

The encryption option for this repository association. It is either owned by Amazon Web Services Key Management Service (KMS) (AWS_OWNED_CMK) or customer managed (CUSTOMER_MANAGED_CMK).

The ID of the Amazon Web Services KMS key that is associated with this respository association.

repositoryAssociation_connectionArn :: Lens' RepositoryAssociation (Maybe Text)

The Amazon Resource Name (ARN) of an Amazon Web Services CodeStar Connections connection. Its format is arn:aws:codestar-connections:region-id:aws-account_id:connection/connection-id. For more information, see Connection in the Amazon Web Services CodeStar Connections API Reference.

repositoryAssociation_stateReason :: Lens' RepositoryAssociation (Maybe Text)

A description of why the repository association is in the current state.

repositoryAssociation_createdTimeStamp :: Lens' RepositoryAssociation (Maybe UTCTime)

The time, in milliseconds since the epoch, when the repository association was created.

repositoryAssociation_lastUpdatedTimeStamp :: Lens' RepositoryAssociation (Maybe UTCTime)

The time, in milliseconds since the epoch, when the repository association was last updated.