Amazonka.CodeGuruReviewer.Types.RepositoryAssociationSummary

Description

Documentation

data RepositoryAssociationSummary

Summary information about a repository association. The ListRepositoryAssociations operation returns a list of RepositoryAssociationSummary objects.

See: newRepositoryAssociationSummary smart constructor.

Constructors

RepositoryAssociationSummary'

Fields

associationArn :: Maybe Text

The Amazon Resource Name (ARN) of the RepositoryAssociation object. You can retrieve this ARN by calling ListRepositoryAssociations .

associationId :: Maybe Text

The repository association ID.

state :: Maybe RepositoryAssociationState

The state of the repository association.

The valid repository association states are:

Associated: The repository association is complete.

Associating: CodeGuru Reviewer is:

Setting up pull request notifications. This is required for pull requests to trigger a CodeGuru Reviewer review.

If your repository ProviderType is GitHub, GitHub Enterprise Server, or Bitbucket, CodeGuru Reviewer creates webhooks in your repository to trigger CodeGuru Reviewer reviews. If you delete these webhooks, reviews of code in your repository cannot be triggered.

Setting up source code access. This is required for CodeGuru Reviewer to securely clone code in your repository.

Failed: The repository failed to associate or disassociate.

Disassociating: CodeGuru Reviewer is removing the repository's pull request notifications and source code access.

Disassociated: CodeGuru Reviewer successfully disassociated the repository. You can create a new association with this repository if you want to review source code in it later. You can control access to code reviews created in an associated repository with tags after it has been disassociated. For more information, see Using tags to control access to associated repositories in the Amazon CodeGuru Reviewer User Guide.

providerType :: Maybe ProviderType

The provider type of the repository association.

owner :: Maybe Text

The owner of the repository. For an Amazon Web Services CodeCommit repository, this is the Amazon Web Services account ID of the account that owns the repository. For a GitHub, GitHub Enterprise Server, or Bitbucket repository, this is the username for the account that owns the repository. For an S3 repository, it can be the username or Amazon Web Services account ID.

name :: Maybe Text

The name of the repository association.

connectionArn :: Maybe Text

The Amazon Resource Name (ARN) of an Amazon Web Services CodeStar Connections connection. Its format is arn:aws:codestar-connections:region-id:aws-account_id:connection/connection-id. For more information, see Connection in the Amazon Web Services CodeStar Connections API Reference.

lastUpdatedTimeStamp :: Maybe POSIX

The time, in milliseconds since the epoch, since the repository association was last updated.

newRepositoryAssociationSummary :: RepositoryAssociationSummary

Create a value of RepositoryAssociationSummary with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:associationArn:RepositoryAssociationSummary', repositoryAssociationSummary_associationArn - The Amazon Resource Name (ARN) of the RepositoryAssociation object. You can retrieve this ARN by calling ListRepositoryAssociations .

$sel:associationId:RepositoryAssociationSummary', repositoryAssociationSummary_associationId - The repository association ID.

$sel:state:RepositoryAssociationSummary', repositoryAssociationSummary_state - The state of the repository association.

The valid repository association states are:

Associated: The repository association is complete.

Associating: CodeGuru Reviewer is:

Setting up pull request notifications. This is required for pull requests to trigger a CodeGuru Reviewer review.

If your repository ProviderType is GitHub, GitHub Enterprise Server, or Bitbucket, CodeGuru Reviewer creates webhooks in your repository to trigger CodeGuru Reviewer reviews. If you delete these webhooks, reviews of code in your repository cannot be triggered.

Setting up source code access. This is required for CodeGuru Reviewer to securely clone code in your repository.

Failed: The repository failed to associate or disassociate.

Disassociating: CodeGuru Reviewer is removing the repository's pull request notifications and source code access.

Disassociated: CodeGuru Reviewer successfully disassociated the repository. You can create a new association with this repository if you want to review source code in it later. You can control access to code reviews created in an associated repository with tags after it has been disassociated. For more information, see Using tags to control access to associated repositories in the Amazon CodeGuru Reviewer User Guide.

$sel:providerType:RepositoryAssociationSummary', repositoryAssociationSummary_providerType - The provider type of the repository association.

$sel:owner:RepositoryAssociationSummary', repositoryAssociationSummary_owner - The owner of the repository. For an Amazon Web Services CodeCommit repository, this is the Amazon Web Services account ID of the account that owns the repository. For a GitHub, GitHub Enterprise Server, or Bitbucket repository, this is the username for the account that owns the repository. For an S3 repository, it can be the username or Amazon Web Services account ID.

$sel:name:RepositoryAssociationSummary', repositoryAssociationSummary_name - The name of the repository association.

$sel:connectionArn:RepositoryAssociationSummary', repositoryAssociationSummary_connectionArn - The Amazon Resource Name (ARN) of an Amazon Web Services CodeStar Connections connection. Its format is arn:aws:codestar-connections:region-id:aws-account_id:connection/connection-id. For more information, see Connection in the Amazon Web Services CodeStar Connections API Reference.

$sel:lastUpdatedTimeStamp:RepositoryAssociationSummary', repositoryAssociationSummary_lastUpdatedTimeStamp - The time, in milliseconds since the epoch, since the repository association was last updated.

repositoryAssociationSummary_associationArn :: Lens' RepositoryAssociationSummary (Maybe Text)

The Amazon Resource Name (ARN) of the RepositoryAssociation object. You can retrieve this ARN by calling ListRepositoryAssociations .

repositoryAssociationSummary_associationId :: Lens' RepositoryAssociationSummary (Maybe Text)

The repository association ID.

repositoryAssociationSummary_state :: Lens' RepositoryAssociationSummary (Maybe RepositoryAssociationState)

The state of the repository association.

The valid repository association states are:

Associated: The repository association is complete.

Associating: CodeGuru Reviewer is:

Setting up pull request notifications. This is required for pull requests to trigger a CodeGuru Reviewer review.

If your repository ProviderType is GitHub, GitHub Enterprise Server, or Bitbucket, CodeGuru Reviewer creates webhooks in your repository to trigger CodeGuru Reviewer reviews. If you delete these webhooks, reviews of code in your repository cannot be triggered.

Setting up source code access. This is required for CodeGuru Reviewer to securely clone code in your repository.

Failed: The repository failed to associate or disassociate.

Disassociating: CodeGuru Reviewer is removing the repository's pull request notifications and source code access.

Disassociated: CodeGuru Reviewer successfully disassociated the repository. You can create a new association with this repository if you want to review source code in it later. You can control access to code reviews created in an associated repository with tags after it has been disassociated. For more information, see Using tags to control access to associated repositories in the Amazon CodeGuru Reviewer User Guide.

repositoryAssociationSummary_providerType :: Lens' RepositoryAssociationSummary (Maybe ProviderType)

The provider type of the repository association.

repositoryAssociationSummary_owner :: Lens' RepositoryAssociationSummary (Maybe Text)

The owner of the repository. For an Amazon Web Services CodeCommit repository, this is the Amazon Web Services account ID of the account that owns the repository. For a GitHub, GitHub Enterprise Server, or Bitbucket repository, this is the username for the account that owns the repository. For an S3 repository, it can be the username or Amazon Web Services account ID.

repositoryAssociationSummary_name :: Lens' RepositoryAssociationSummary (Maybe Text)

The name of the repository association.

repositoryAssociationSummary_connectionArn :: Lens' RepositoryAssociationSummary (Maybe Text)

The Amazon Resource Name (ARN) of an Amazon Web Services CodeStar Connections connection. Its format is arn:aws:codestar-connections:region-id:aws-account_id:connection/connection-id. For more information, see Connection in the Amazon Web Services CodeStar Connections API Reference.

repositoryAssociationSummary_lastUpdatedTimeStamp :: Lens' RepositoryAssociationSummary (Maybe UTCTime)

The time, in milliseconds since the epoch, since the repository association was last updated.