Amazonka.CodeGuruReviewer.Types.RuleMetadata

Description

Documentation

data RuleMetadata

Metadata about a rule. Rule metadata includes an ID, a name, a list of tags, and a short and long description. CodeGuru Reviewer uses rules to analyze code. A rule's recommendation is included in analysis results if code is detected that violates the rule.

See: newRuleMetadata smart constructor.

Constructors

RuleMetadata'

Fields

longDescription :: Maybe Text

A long description of the rule.

ruleTags :: Maybe (NonEmpty Text)

Tags that are associated with the rule.

ruleId :: Maybe Text

The ID of the rule.

ruleName :: Maybe Text

The name of the rule.

shortDescription :: Maybe Text

A short description of the rule.

newRuleMetadata :: RuleMetadata

Create a value of RuleMetadata with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:longDescription:RuleMetadata', ruleMetadata_longDescription - A long description of the rule.

$sel:ruleTags:RuleMetadata', ruleMetadata_ruleTags - Tags that are associated with the rule.

$sel:ruleId:RuleMetadata', ruleMetadata_ruleId - The ID of the rule.

$sel:ruleName:RuleMetadata', ruleMetadata_ruleName - The name of the rule.

$sel:shortDescription:RuleMetadata', ruleMetadata_shortDescription - A short description of the rule.

ruleMetadata_longDescription :: Lens' RuleMetadata (Maybe Text)

A long description of the rule.

ruleMetadata_ruleTags :: Lens' RuleMetadata (Maybe (NonEmpty Text))

Tags that are associated with the rule.

ruleMetadata_ruleId :: Lens' RuleMetadata (Maybe Text)

The ID of the rule.

ruleMetadata_ruleName :: Lens' RuleMetadata (Maybe Text)

The name of the rule.

ruleMetadata_shortDescription :: Lens' RuleMetadata (Maybe Text)

A short description of the rule.