Amazonka.CodeGuruReviewer.Types.ThirdPartySourceRepository

Description

Documentation

data ThirdPartySourceRepository

Information about a third-party source repository connected to CodeGuru Reviewer.

See: newThirdPartySourceRepository smart constructor.

Constructors

ThirdPartySourceRepository'

Fields

name :: Text

The name of the third party source repository.

connectionArn :: Text

The Amazon Resource Name (ARN) of an Amazon Web Services CodeStar Connections connection. Its format is arn:aws:codestar-connections:region-id:aws-account_id:connection/connection-id. For more information, see Connection in the Amazon Web Services CodeStar Connections API Reference.

owner :: Text

The owner of the repository. For a GitHub, GitHub Enterprise, or Bitbucket repository, this is the username for the account that owns the repository. For an S3 repository, this can be the username or Amazon Web Services account ID.

newThirdPartySourceRepository

Arguments

:: Text
$sel:name:ThirdPartySourceRepository'

-> Text
$sel:connectionArn:ThirdPartySourceRepository'

-> Text
$sel:owner:ThirdPartySourceRepository'

-> ThirdPartySourceRepository

Create a value of ThirdPartySourceRepository with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:name:ThirdPartySourceRepository', thirdPartySourceRepository_name - The name of the third party source repository.

$sel:connectionArn:ThirdPartySourceRepository', thirdPartySourceRepository_connectionArn - The Amazon Resource Name (ARN) of an Amazon Web Services CodeStar Connections connection. Its format is arn:aws:codestar-connections:region-id:aws-account_id:connection/connection-id. For more information, see Connection in the Amazon Web Services CodeStar Connections API Reference.

$sel:owner:ThirdPartySourceRepository', thirdPartySourceRepository_owner - The owner of the repository. For a GitHub, GitHub Enterprise, or Bitbucket repository, this is the username for the account that owns the repository. For an S3 repository, this can be the username or Amazon Web Services account ID.

thirdPartySourceRepository_name :: Lens' ThirdPartySourceRepository Text

The name of the third party source repository.

thirdPartySourceRepository_connectionArn :: Lens' ThirdPartySourceRepository Text

The Amazon Resource Name (ARN) of an Amazon Web Services CodeStar Connections connection. Its format is arn:aws:codestar-connections:region-id:aws-account_id:connection/connection-id. For more information, see Connection in the Amazon Web Services CodeStar Connections API Reference.

thirdPartySourceRepository_owner :: Lens' ThirdPartySourceRepository Text

The owner of the repository. For a GitHub, GitHub Enterprise, or Bitbucket repository, this is the username for the account that owns the repository. For an S3 repository, this can be the username or Amazon Web Services account ID.