Amazonka.CodePipeline.AcknowledgeThirdPartyJob

Description

Confirms a job worker has received the specified job. Used for partner actions only.

Creating a Request

data AcknowledgeThirdPartyJob

Represents the input of an AcknowledgeThirdPartyJob action.

See: newAcknowledgeThirdPartyJob smart constructor.

Constructors

AcknowledgeThirdPartyJob'

Fields

jobId :: Text

The unique system-generated ID of the job.

nonce :: Text

A system-generated random number that AWS CodePipeline uses to ensure that the job is being worked on by only one job worker. Get this number from the response to a GetThirdPartyJobDetails request.

clientToken :: Text

The clientToken portion of the clientId and clientToken pair used to verify that the calling entity is allowed access to the job and its details.

newAcknowledgeThirdPartyJob

Arguments

:: Text
$sel:jobId:AcknowledgeThirdPartyJob'

-> Text
$sel:nonce:AcknowledgeThirdPartyJob'

-> Text
$sel:clientToken:AcknowledgeThirdPartyJob'

-> AcknowledgeThirdPartyJob

Create a value of AcknowledgeThirdPartyJob with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:jobId:AcknowledgeThirdPartyJob', acknowledgeThirdPartyJob_jobId - The unique system-generated ID of the job.

$sel:nonce:AcknowledgeThirdPartyJob', acknowledgeThirdPartyJob_nonce - A system-generated random number that AWS CodePipeline uses to ensure that the job is being worked on by only one job worker. Get this number from the response to a GetThirdPartyJobDetails request.

$sel:clientToken:AcknowledgeThirdPartyJob', acknowledgeThirdPartyJob_clientToken - The clientToken portion of the clientId and clientToken pair used to verify that the calling entity is allowed access to the job and its details.

Request Lenses

acknowledgeThirdPartyJob_jobId :: Lens' AcknowledgeThirdPartyJob Text

The unique system-generated ID of the job.

acknowledgeThirdPartyJob_nonce :: Lens' AcknowledgeThirdPartyJob Text

A system-generated random number that AWS CodePipeline uses to ensure that the job is being worked on by only one job worker. Get this number from the response to a GetThirdPartyJobDetails request.

acknowledgeThirdPartyJob_clientToken :: Lens' AcknowledgeThirdPartyJob Text

The clientToken portion of the clientId and clientToken pair used to verify that the calling entity is allowed access to the job and its details.

Destructuring the Response

data AcknowledgeThirdPartyJobResponse

Represents the output of an AcknowledgeThirdPartyJob action.

See: newAcknowledgeThirdPartyJobResponse smart constructor.

Constructors

AcknowledgeThirdPartyJobResponse'

Fields

status :: Maybe JobStatus

The status information for the third party job, if any.

httpStatus :: Int

The response's http status code.

newAcknowledgeThirdPartyJobResponse

Arguments

:: Int
$sel:httpStatus:AcknowledgeThirdPartyJobResponse'

-> AcknowledgeThirdPartyJobResponse

Create a value of AcknowledgeThirdPartyJobResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:status:AcknowledgeThirdPartyJobResponse', acknowledgeThirdPartyJobResponse_status - The status information for the third party job, if any.

$sel:httpStatus:AcknowledgeThirdPartyJobResponse', acknowledgeThirdPartyJobResponse_httpStatus - The response's http status code.

Response Lenses

acknowledgeThirdPartyJobResponse_status :: Lens' AcknowledgeThirdPartyJobResponse (Maybe JobStatus)

The status information for the third party job, if any.

acknowledgeThirdPartyJobResponse_httpStatus :: Lens' AcknowledgeThirdPartyJobResponse Int

The response's http status code.