Amazonka.CodePipeline.GetJobDetails

Description

Returns information about a job. Used for custom actions only.

When this API is called, AWS CodePipeline returns temporary credentials for the S3 bucket used to store artifacts for the pipeline, if the action requires access to that S3 bucket for input or output artifacts. This API also returns any secret values defined for the action.

Creating a Request

data GetJobDetails

Represents the input of a GetJobDetails action.

See: newGetJobDetails smart constructor.

Constructors

GetJobDetails'

Fields

jobId :: Text

The unique system-generated ID for the job.

newGetJobDetails

Arguments

:: Text
$sel:jobId:GetJobDetails'

-> GetJobDetails

Create a value of GetJobDetails with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:jobId:GetJobDetails', getJobDetails_jobId - The unique system-generated ID for the job.

Request Lenses

getJobDetails_jobId :: Lens' GetJobDetails Text

The unique system-generated ID for the job.

Destructuring the Response

data GetJobDetailsResponse

Represents the output of a GetJobDetails action.

See: newGetJobDetailsResponse smart constructor.

Constructors

GetJobDetailsResponse'

Fields

jobDetails :: Maybe JobDetails

The details of the job.

If AWSSessionCredentials is used, a long-running job can call GetJobDetails again to obtain new credentials.

httpStatus :: Int

The response's http status code.

newGetJobDetailsResponse

Arguments

:: Int
$sel:httpStatus:GetJobDetailsResponse'

-> GetJobDetailsResponse

Create a value of GetJobDetailsResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:jobDetails:GetJobDetailsResponse', getJobDetailsResponse_jobDetails - The details of the job.

If AWSSessionCredentials is used, a long-running job can call GetJobDetails again to obtain new credentials.

$sel:httpStatus:GetJobDetailsResponse', getJobDetailsResponse_httpStatus - The response's http status code.

Response Lenses

getJobDetailsResponse_jobDetails :: Lens' GetJobDetailsResponse (Maybe JobDetails)

The details of the job.

If AWSSessionCredentials is used, a long-running job can call GetJobDetails again to obtain new credentials.

getJobDetailsResponse_httpStatus :: Lens' GetJobDetailsResponse Int

The response's http status code.