Amazonka.CodePipeline.GetPipelineState

Description

Returns information about the state of a pipeline, including the stages and actions.

Values returned in the revisionId and revisionUrl fields indicate the source revision information, such as the commit ID, for the current state.

Creating a Request

data GetPipelineState

Represents the input of a GetPipelineState action.

See: newGetPipelineState smart constructor.

Constructors

GetPipelineState'

Fields

name :: Text

The name of the pipeline about which you want to get information.

newGetPipelineState

Arguments

:: Text
$sel:name:GetPipelineState'

-> GetPipelineState

Create a value of GetPipelineState with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:name:GetPipelineState', getPipelineState_name - The name of the pipeline about which you want to get information.

Request Lenses

getPipelineState_name :: Lens' GetPipelineState Text

The name of the pipeline about which you want to get information.

Destructuring the Response

data GetPipelineStateResponse

Represents the output of a GetPipelineState action.

See: newGetPipelineStateResponse smart constructor.

Constructors

GetPipelineStateResponse'

Fields

pipelineName :: Maybe Text

The name of the pipeline for which you want to get the state.

created :: Maybe POSIX

The date and time the pipeline was created, in timestamp format.

stageStates :: Maybe [StageState]

A list of the pipeline stage output information, including stage name, state, most recent run details, whether the stage is disabled, and other data.

pipelineVersion :: Maybe Natural

The version number of the pipeline.

A newly created pipeline is always assigned a version number of 1.

updated :: Maybe POSIX

The date and time the pipeline was last updated, in timestamp format.

httpStatus :: Int

The response's http status code.

newGetPipelineStateResponse

Arguments

:: Int
$sel:httpStatus:GetPipelineStateResponse'

-> GetPipelineStateResponse

Create a value of GetPipelineStateResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:pipelineName:GetPipelineStateResponse', getPipelineStateResponse_pipelineName - The name of the pipeline for which you want to get the state.

$sel:created:GetPipelineStateResponse', getPipelineStateResponse_created - The date and time the pipeline was created, in timestamp format.

$sel:stageStates:GetPipelineStateResponse', getPipelineStateResponse_stageStates - A list of the pipeline stage output information, including stage name, state, most recent run details, whether the stage is disabled, and other data.

$sel:pipelineVersion:GetPipelineStateResponse', getPipelineStateResponse_pipelineVersion - The version number of the pipeline.

A newly created pipeline is always assigned a version number of 1.

$sel:updated:GetPipelineStateResponse', getPipelineStateResponse_updated - The date and time the pipeline was last updated, in timestamp format.

$sel:httpStatus:GetPipelineStateResponse', getPipelineStateResponse_httpStatus - The response's http status code.

Response Lenses

getPipelineStateResponse_pipelineName :: Lens' GetPipelineStateResponse (Maybe Text)

The name of the pipeline for which you want to get the state.

getPipelineStateResponse_created :: Lens' GetPipelineStateResponse (Maybe UTCTime)

The date and time the pipeline was created, in timestamp format.

getPipelineStateResponse_stageStates :: Lens' GetPipelineStateResponse (Maybe [StageState])

A list of the pipeline stage output information, including stage name, state, most recent run details, whether the stage is disabled, and other data.

getPipelineStateResponse_pipelineVersion :: Lens' GetPipelineStateResponse (Maybe Natural)

The version number of the pipeline.

A newly created pipeline is always assigned a version number of 1.

getPipelineStateResponse_updated :: Lens' GetPipelineStateResponse (Maybe UTCTime)

The date and time the pipeline was last updated, in timestamp format.

getPipelineStateResponse_httpStatus :: Lens' GetPipelineStateResponse Int

The response's http status code.