Amazonka.CodePipeline.ListActionTypes

Description

Gets a summary of all AWS CodePipeline action types associated with your account.

This operation returns paginated results.

Creating a Request

data ListActionTypes

Represents the input of a ListActionTypes action.

See: newListActionTypes smart constructor.

Constructors

ListActionTypes'

Fields

actionOwnerFilter :: Maybe ActionOwner

Filters the list of action types to those created by a specified entity.

regionFilter :: Maybe Text

The Region to filter on for the list of action types.

nextToken :: Maybe Text

An identifier that was returned from the previous list action types call, which can be used to return the next set of action types in the list.

newListActionTypes :: ListActionTypes

Create a value of ListActionTypes with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:actionOwnerFilter:ListActionTypes', listActionTypes_actionOwnerFilter - Filters the list of action types to those created by a specified entity.

$sel:regionFilter:ListActionTypes', listActionTypes_regionFilter - The Region to filter on for the list of action types.

$sel:nextToken:ListActionTypes', listActionTypes_nextToken - An identifier that was returned from the previous list action types call, which can be used to return the next set of action types in the list.

Request Lenses

listActionTypes_actionOwnerFilter :: Lens' ListActionTypes (Maybe ActionOwner)

Filters the list of action types to those created by a specified entity.

listActionTypes_regionFilter :: Lens' ListActionTypes (Maybe Text)

The Region to filter on for the list of action types.

listActionTypes_nextToken :: Lens' ListActionTypes (Maybe Text)

An identifier that was returned from the previous list action types call, which can be used to return the next set of action types in the list.

Destructuring the Response

data ListActionTypesResponse

Represents the output of a ListActionTypes action.

See: newListActionTypesResponse smart constructor.

Constructors

ListActionTypesResponse'

Fields

nextToken :: Maybe Text

If the amount of returned information is significantly large, an identifier is also returned. It can be used in a subsequent list action types call to return the next set of action types in the list.

httpStatus :: Int

The response's http status code.

actionTypes :: [ActionType]

Provides details of the action types.

newListActionTypesResponse

Arguments

:: Int
$sel:httpStatus:ListActionTypesResponse'

-> ListActionTypesResponse

Create a value of ListActionTypesResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:nextToken:ListActionTypes', listActionTypesResponse_nextToken - If the amount of returned information is significantly large, an identifier is also returned. It can be used in a subsequent list action types call to return the next set of action types in the list.

$sel:httpStatus:ListActionTypesResponse', listActionTypesResponse_httpStatus - The response's http status code.

$sel:actionTypes:ListActionTypesResponse', listActionTypesResponse_actionTypes - Provides details of the action types.

Response Lenses

listActionTypesResponse_nextToken :: Lens' ListActionTypesResponse (Maybe Text)

If the amount of returned information is significantly large, an identifier is also returned. It can be used in a subsequent list action types call to return the next set of action types in the list.

listActionTypesResponse_httpStatus :: Lens' ListActionTypesResponse Int

The response's http status code.

listActionTypesResponse_actionTypes :: Lens' ListActionTypesResponse [ActionType]

Provides details of the action types.