Amazonka.CodePipeline.PollForJobs

Description

Returns information about any jobs for AWS CodePipeline to act on. PollForJobs is valid only for action types with "Custom" in the owner field. If the action type contains "AWS" or "ThirdParty" in the owner field, the PollForJobs action returns an error.

When this API is called, AWS CodePipeline returns temporary credentials for the S3 bucket used to store artifacts for the pipeline, if the action requires access to that S3 bucket for input or output artifacts. This API also returns any secret values defined for the action.

Creating a Request

data PollForJobs

Represents the input of a PollForJobs action.

See: newPollForJobs smart constructor.

Constructors

PollForJobs'

Fields

maxBatchSize :: Maybe Natural

The maximum number of jobs to return in a poll for jobs call.

queryParam :: Maybe (HashMap Text Text)

A map of property names and values. For an action type with no queryable properties, this value must be null or an empty map. For an action type with a queryable property, you must supply that property as a key in the map. Only jobs whose action configuration matches the mapped value are returned.

actionTypeId :: ActionTypeId

Represents information about an action type.

newPollForJobs

Arguments

:: ActionTypeId
$sel:actionTypeId:PollForJobs'

-> PollForJobs

Create a value of PollForJobs with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:maxBatchSize:PollForJobs', pollForJobs_maxBatchSize - The maximum number of jobs to return in a poll for jobs call.

$sel:queryParam:PollForJobs', pollForJobs_queryParam - A map of property names and values. For an action type with no queryable properties, this value must be null or an empty map. For an action type with a queryable property, you must supply that property as a key in the map. Only jobs whose action configuration matches the mapped value are returned.

$sel:actionTypeId:PollForJobs', pollForJobs_actionTypeId - Represents information about an action type.

Request Lenses

pollForJobs_maxBatchSize :: Lens' PollForJobs (Maybe Natural)

The maximum number of jobs to return in a poll for jobs call.

pollForJobs_queryParam :: Lens' PollForJobs (Maybe (HashMap Text Text))

A map of property names and values. For an action type with no queryable properties, this value must be null or an empty map. For an action type with a queryable property, you must supply that property as a key in the map. Only jobs whose action configuration matches the mapped value are returned.

pollForJobs_actionTypeId :: Lens' PollForJobs ActionTypeId

Represents information about an action type.

Destructuring the Response

data PollForJobsResponse

Represents the output of a PollForJobs action.

See: newPollForJobsResponse smart constructor.

Constructors

PollForJobsResponse'

Fields

jobs :: Maybe [Job]

Information about the jobs to take action on.

httpStatus :: Int

The response's http status code.

newPollForJobsResponse

Arguments

:: Int
$sel:httpStatus:PollForJobsResponse'

-> PollForJobsResponse

Create a value of PollForJobsResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:jobs:PollForJobsResponse', pollForJobsResponse_jobs - Information about the jobs to take action on.

$sel:httpStatus:PollForJobsResponse', pollForJobsResponse_httpStatus - The response's http status code.

Response Lenses

pollForJobsResponse_jobs :: Lens' PollForJobsResponse (Maybe [Job])

Information about the jobs to take action on.

pollForJobsResponse_httpStatus :: Lens' PollForJobsResponse Int

The response's http status code.