Amazonka.CodePipeline.PollForThirdPartyJobs

Description

Determines whether there are any third party jobs for a job worker to act on. Used for partner actions only.

When this API is called, AWS CodePipeline returns temporary credentials for the S3 bucket used to store artifacts for the pipeline, if the action requires access to that S3 bucket for input or output artifacts.

Creating a Request

data PollForThirdPartyJobs

Represents the input of a PollForThirdPartyJobs action.

See: newPollForThirdPartyJobs smart constructor.

Constructors

PollForThirdPartyJobs'

Fields

maxBatchSize :: Maybe Natural

The maximum number of jobs to return in a poll for jobs call.

actionTypeId :: ActionTypeId

Represents information about an action type.

newPollForThirdPartyJobs

Arguments

:: ActionTypeId
$sel:actionTypeId:PollForThirdPartyJobs'

-> PollForThirdPartyJobs

Create a value of PollForThirdPartyJobs with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:maxBatchSize:PollForThirdPartyJobs', pollForThirdPartyJobs_maxBatchSize - The maximum number of jobs to return in a poll for jobs call.

$sel:actionTypeId:PollForThirdPartyJobs', pollForThirdPartyJobs_actionTypeId - Represents information about an action type.

Request Lenses

pollForThirdPartyJobs_maxBatchSize :: Lens' PollForThirdPartyJobs (Maybe Natural)

The maximum number of jobs to return in a poll for jobs call.

pollForThirdPartyJobs_actionTypeId :: Lens' PollForThirdPartyJobs ActionTypeId

Represents information about an action type.

Destructuring the Response

data PollForThirdPartyJobsResponse

Represents the output of a PollForThirdPartyJobs action.

See: newPollForThirdPartyJobsResponse smart constructor.

Constructors

PollForThirdPartyJobsResponse'

Fields

jobs :: Maybe [ThirdPartyJob]

Information about the jobs to take action on.

httpStatus :: Int

The response's http status code.

newPollForThirdPartyJobsResponse

Arguments

:: Int
$sel:httpStatus:PollForThirdPartyJobsResponse'

-> PollForThirdPartyJobsResponse

Create a value of PollForThirdPartyJobsResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:jobs:PollForThirdPartyJobsResponse', pollForThirdPartyJobsResponse_jobs - Information about the jobs to take action on.

$sel:httpStatus:PollForThirdPartyJobsResponse', pollForThirdPartyJobsResponse_httpStatus - The response's http status code.

Response Lenses

pollForThirdPartyJobsResponse_jobs :: Lens' PollForThirdPartyJobsResponse (Maybe [ThirdPartyJob])

Information about the jobs to take action on.

pollForThirdPartyJobsResponse_httpStatus :: Lens' PollForThirdPartyJobsResponse Int

The response's http status code.