Amazonka.CodePipeline.StopPipelineExecution

Description

Stops the specified pipeline execution. You choose to either stop the pipeline execution by completing in-progress actions without starting subsequent actions, or by abandoning in-progress actions. While completing or abandoning in-progress actions, the pipeline execution is in a Stopping state. After all in-progress actions are completed or abandoned, the pipeline execution is in a Stopped state.

Creating a Request

data StopPipelineExecution

See: newStopPipelineExecution smart constructor.

Constructors

StopPipelineExecution'

Fields

abandon :: Maybe Bool

Use this option to stop the pipeline execution by abandoning, rather than finishing, in-progress actions.

This option can lead to failed or out-of-sequence tasks.

reason :: Maybe Text

Use this option to enter comments, such as the reason the pipeline was stopped.

pipelineName :: Text

The name of the pipeline to stop.

pipelineExecutionId :: Text

The ID of the pipeline execution to be stopped in the current stage. Use the GetPipelineState action to retrieve the current pipelineExecutionId.

newStopPipelineExecution

Arguments

:: Text
$sel:pipelineName:StopPipelineExecution'

-> Text
$sel:pipelineExecutionId:StopPipelineExecution'

-> StopPipelineExecution

Create a value of StopPipelineExecution with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:abandon:StopPipelineExecution', stopPipelineExecution_abandon - Use this option to stop the pipeline execution by abandoning, rather than finishing, in-progress actions.

This option can lead to failed or out-of-sequence tasks.

$sel:reason:StopPipelineExecution', stopPipelineExecution_reason - Use this option to enter comments, such as the reason the pipeline was stopped.

$sel:pipelineName:StopPipelineExecution', stopPipelineExecution_pipelineName - The name of the pipeline to stop.

$sel:pipelineExecutionId:StopPipelineExecution', stopPipelineExecution_pipelineExecutionId - The ID of the pipeline execution to be stopped in the current stage. Use the GetPipelineState action to retrieve the current pipelineExecutionId.

Request Lenses

stopPipelineExecution_abandon :: Lens' StopPipelineExecution (Maybe Bool)

Use this option to stop the pipeline execution by abandoning, rather than finishing, in-progress actions.

This option can lead to failed or out-of-sequence tasks.

stopPipelineExecution_reason :: Lens' StopPipelineExecution (Maybe Text)

Use this option to enter comments, such as the reason the pipeline was stopped.

stopPipelineExecution_pipelineName :: Lens' StopPipelineExecution Text

The name of the pipeline to stop.

stopPipelineExecution_pipelineExecutionId :: Lens' StopPipelineExecution Text

The ID of the pipeline execution to be stopped in the current stage. Use the GetPipelineState action to retrieve the current pipelineExecutionId.

Destructuring the Response

data StopPipelineExecutionResponse

See: newStopPipelineExecutionResponse smart constructor.

Constructors

StopPipelineExecutionResponse'

Fields

pipelineExecutionId :: Maybe Text

The unique system-generated ID of the pipeline execution that was stopped.

httpStatus :: Int

The response's http status code.

newStopPipelineExecutionResponse

Arguments

:: Int
$sel:httpStatus:StopPipelineExecutionResponse'

-> StopPipelineExecutionResponse

Create a value of StopPipelineExecutionResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:pipelineExecutionId:StopPipelineExecution', stopPipelineExecutionResponse_pipelineExecutionId - The unique system-generated ID of the pipeline execution that was stopped.

$sel:httpStatus:StopPipelineExecutionResponse', stopPipelineExecutionResponse_httpStatus - The response's http status code.

Response Lenses

stopPipelineExecutionResponse_pipelineExecutionId :: Lens' StopPipelineExecutionResponse (Maybe Text)

The unique system-generated ID of the pipeline execution that was stopped.

stopPipelineExecutionResponse_httpStatus :: Lens' StopPipelineExecutionResponse Int

The response's http status code.