Amazonka.CodePipeline.Types.ActionType

Description

Documentation

data ActionType

Returns information about the details of an action type.

See: newActionType smart constructor.

Constructors

ActionType'

Fields

settings :: Maybe ActionTypeSettings

The settings for the action type.

actionConfigurationProperties :: Maybe [ActionConfigurationProperty]

The configuration properties for the action type.

id :: ActionTypeId

Represents information about an action type.

inputArtifactDetails :: ArtifactDetails

The details of the input artifact for the action, such as its commit ID.

outputArtifactDetails :: ArtifactDetails

The details of the output artifact of the action, such as its commit ID.

newActionType

Arguments

:: ActionTypeId
$sel:id:ActionType'

-> ArtifactDetails
$sel:inputArtifactDetails:ActionType'

-> ArtifactDetails
$sel:outputArtifactDetails:ActionType'

-> ActionType

Create a value of ActionType with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:settings:ActionType', actionType_settings - The settings for the action type.

$sel:actionConfigurationProperties:ActionType', actionType_actionConfigurationProperties - The configuration properties for the action type.

$sel:id:ActionType', actionType_id - Represents information about an action type.

$sel:inputArtifactDetails:ActionType', actionType_inputArtifactDetails - The details of the input artifact for the action, such as its commit ID.

$sel:outputArtifactDetails:ActionType', actionType_outputArtifactDetails - The details of the output artifact of the action, such as its commit ID.

actionType_settings :: Lens' ActionType (Maybe ActionTypeSettings)

The settings for the action type.

actionType_actionConfigurationProperties :: Lens' ActionType (Maybe [ActionConfigurationProperty])

The configuration properties for the action type.

actionType_id :: Lens' ActionType ActionTypeId

Represents information about an action type.

actionType_inputArtifactDetails :: Lens' ActionType ArtifactDetails

The details of the input artifact for the action, such as its commit ID.

actionType_outputArtifactDetails :: Lens' ActionType ArtifactDetails

The details of the output artifact of the action, such as its commit ID.