Amazonka.CodePipeline.Types.ActionTypeSettings

Description

Documentation

data ActionTypeSettings

Returns information about the settings for an action type.

See: newActionTypeSettings smart constructor.

Constructors

ActionTypeSettings'

Fields

thirdPartyConfigurationUrl :: Maybe Text

The URL of a sign-up page where users can sign up for an external service and perform initial configuration of the action provided by that service.

executionUrlTemplate :: Maybe Text

The URL returned to the AWS CodePipeline console that contains a link to the top-level landing page for the external system, such as the console page for AWS CodeDeploy. This link is shown on the pipeline view page in the AWS CodePipeline console and provides a link to the execution entity of the external action.

revisionUrlTemplate :: Maybe Text

The URL returned to the AWS CodePipeline console that contains a link to the page where customers can update or change the configuration of the external action.

entityUrlTemplate :: Maybe Text

The URL returned to the AWS CodePipeline console that provides a deep link to the resources of the external system, such as the configuration page for an AWS CodeDeploy deployment group. This link is provided as part of the action display in the pipeline.

newActionTypeSettings :: ActionTypeSettings

Create a value of ActionTypeSettings with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:thirdPartyConfigurationUrl:ActionTypeSettings', actionTypeSettings_thirdPartyConfigurationUrl - The URL of a sign-up page where users can sign up for an external service and perform initial configuration of the action provided by that service.

$sel:executionUrlTemplate:ActionTypeSettings', actionTypeSettings_executionUrlTemplate - The URL returned to the AWS CodePipeline console that contains a link to the top-level landing page for the external system, such as the console page for AWS CodeDeploy. This link is shown on the pipeline view page in the AWS CodePipeline console and provides a link to the execution entity of the external action.

$sel:revisionUrlTemplate:ActionTypeSettings', actionTypeSettings_revisionUrlTemplate - The URL returned to the AWS CodePipeline console that contains a link to the page where customers can update or change the configuration of the external action.

$sel:entityUrlTemplate:ActionTypeSettings', actionTypeSettings_entityUrlTemplate - The URL returned to the AWS CodePipeline console that provides a deep link to the resources of the external system, such as the configuration page for an AWS CodeDeploy deployment group. This link is provided as part of the action display in the pipeline.

actionTypeSettings_thirdPartyConfigurationUrl :: Lens' ActionTypeSettings (Maybe Text)

The URL of a sign-up page where users can sign up for an external service and perform initial configuration of the action provided by that service.

actionTypeSettings_executionUrlTemplate :: Lens' ActionTypeSettings (Maybe Text)

The URL returned to the AWS CodePipeline console that contains a link to the top-level landing page for the external system, such as the console page for AWS CodeDeploy. This link is shown on the pipeline view page in the AWS CodePipeline console and provides a link to the execution entity of the external action.

actionTypeSettings_revisionUrlTemplate :: Lens' ActionTypeSettings (Maybe Text)

The URL returned to the AWS CodePipeline console that contains a link to the page where customers can update or change the configuration of the external action.

actionTypeSettings_entityUrlTemplate :: Lens' ActionTypeSettings (Maybe Text)

The URL returned to the AWS CodePipeline console that provides a deep link to the resources of the external system, such as the configuration page for an AWS CodeDeploy deployment group. This link is provided as part of the action display in the pipeline.