Amazonka.CodePipeline.Types.ArtifactStore

Description

Documentation

data ArtifactStore

The S3 bucket where artifacts for the pipeline are stored.

You must include either artifactStore or artifactStores in your pipeline, but you cannot use both. If you create a cross-region action in your pipeline, you must use artifactStores.

See: newArtifactStore smart constructor.

Constructors

ArtifactStore'

Fields

encryptionKey :: Maybe EncryptionKey

The encryption key used to encrypt the data in the artifact store, such as an AWS Key Management Service (AWS KMS) key. If this is undefined, the default key for Amazon S3 is used.

type' :: ArtifactStoreType

The type of the artifact store, such as S3.

location :: Text

The S3 bucket used for storing the artifacts for a pipeline. You can specify the name of an S3 bucket but not a folder in the bucket. A folder to contain the pipeline artifacts is created for you based on the name of the pipeline. You can use any S3 bucket in the same AWS Region as the pipeline to store your pipeline artifacts.

newArtifactStore

Arguments

:: ArtifactStoreType
$sel:type':ArtifactStore'

-> Text
$sel:location:ArtifactStore'

-> ArtifactStore

Create a value of ArtifactStore with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:encryptionKey:ArtifactStore', artifactStore_encryptionKey - The encryption key used to encrypt the data in the artifact store, such as an AWS Key Management Service (AWS KMS) key. If this is undefined, the default key for Amazon S3 is used.

$sel:type':ArtifactStore', artifactStore_type - The type of the artifact store, such as S3.

$sel:location:ArtifactStore', artifactStore_location - The S3 bucket used for storing the artifacts for a pipeline. You can specify the name of an S3 bucket but not a folder in the bucket. A folder to contain the pipeline artifacts is created for you based on the name of the pipeline. You can use any S3 bucket in the same AWS Region as the pipeline to store your pipeline artifacts.

artifactStore_encryptionKey :: Lens' ArtifactStore (Maybe EncryptionKey)

The encryption key used to encrypt the data in the artifact store, such as an AWS Key Management Service (AWS KMS) key. If this is undefined, the default key for Amazon S3 is used.

artifactStore_type :: Lens' ArtifactStore ArtifactStoreType

The type of the artifact store, such as S3.

artifactStore_location :: Lens' ArtifactStore Text

The S3 bucket used for storing the artifacts for a pipeline. You can specify the name of an S3 bucket but not a folder in the bucket. A folder to contain the pipeline artifacts is created for you based on the name of the pipeline. You can use any S3 bucket in the same AWS Region as the pipeline to store your pipeline artifacts.