Amazonka.CodePipeline.Types.SourceRevision

Description

Documentation

data SourceRevision

Information about the version (or revision) of a source artifact that initiated a pipeline execution.

See: newSourceRevision smart constructor.

Constructors

SourceRevision'

Fields

revisionSummary :: Maybe Text

Summary information about the most recent revision of the artifact. For GitHub and AWS CodeCommit repositories, the commit message. For Amazon S3 buckets or actions, the user-provided content of a codepipeline-artifact-revision-summary key specified in the object metadata.

revisionUrl :: Maybe Text

The commit ID for the artifact revision. For artifacts stored in GitHub or AWS CodeCommit repositories, the commit ID is linked to a commit details page.

revisionId :: Maybe Text

The system-generated unique ID that identifies the revision number of the artifact.

actionName :: Text

The name of the action that processed the revision to the source artifact.

newSourceRevision

Arguments

:: Text
$sel:actionName:SourceRevision'

-> SourceRevision

Create a value of SourceRevision with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:revisionSummary:SourceRevision', sourceRevision_revisionSummary - Summary information about the most recent revision of the artifact. For GitHub and AWS CodeCommit repositories, the commit message. For Amazon S3 buckets or actions, the user-provided content of a codepipeline-artifact-revision-summary key specified in the object metadata.

$sel:revisionUrl:SourceRevision', sourceRevision_revisionUrl - The commit ID for the artifact revision. For artifacts stored in GitHub or AWS CodeCommit repositories, the commit ID is linked to a commit details page.

$sel:revisionId:SourceRevision', sourceRevision_revisionId - The system-generated unique ID that identifies the revision number of the artifact.

$sel:actionName:SourceRevision', sourceRevision_actionName - The name of the action that processed the revision to the source artifact.

sourceRevision_revisionSummary :: Lens' SourceRevision (Maybe Text)

Summary information about the most recent revision of the artifact. For GitHub and AWS CodeCommit repositories, the commit message. For Amazon S3 buckets or actions, the user-provided content of a codepipeline-artifact-revision-summary key specified in the object metadata.

sourceRevision_revisionUrl :: Lens' SourceRevision (Maybe Text)

The commit ID for the artifact revision. For artifacts stored in GitHub or AWS CodeCommit repositories, the commit ID is linked to a commit details page.

sourceRevision_revisionId :: Lens' SourceRevision (Maybe Text)

The system-generated unique ID that identifies the revision number of the artifact.

sourceRevision_actionName :: Lens' SourceRevision Text

The name of the action that processed the revision to the source artifact.