Amazonka.CodeStar.ListTagsForProject

Description

Gets the tags for a project.

Creating a Request

data ListTagsForProject

See: newListTagsForProject smart constructor.

Constructors

ListTagsForProject'

Fields

nextToken :: Maybe Text

Reserved for future use.

maxResults :: Maybe Natural

Reserved for future use.

id :: Text

The ID of the project to get tags for.

newListTagsForProject

Arguments

:: Text
$sel:id:ListTagsForProject'

-> ListTagsForProject

Create a value of ListTagsForProject with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:nextToken:ListTagsForProject', listTagsForProject_nextToken - Reserved for future use.

$sel:maxResults:ListTagsForProject', listTagsForProject_maxResults - Reserved for future use.

$sel:id:ListTagsForProject', listTagsForProject_id - The ID of the project to get tags for.

Request Lenses

listTagsForProject_nextToken :: Lens' ListTagsForProject (Maybe Text)

Reserved for future use.

listTagsForProject_maxResults :: Lens' ListTagsForProject (Maybe Natural)

Reserved for future use.

listTagsForProject_id :: Lens' ListTagsForProject Text

The ID of the project to get tags for.

Destructuring the Response

data ListTagsForProjectResponse

See: newListTagsForProjectResponse smart constructor.

Constructors

ListTagsForProjectResponse'

Fields

nextToken :: Maybe Text

Reserved for future use.

tags :: Maybe (HashMap Text Text)

The tags for the project.

httpStatus :: Int

The response's http status code.

newListTagsForProjectResponse

Arguments

:: Int
$sel:httpStatus:ListTagsForProjectResponse'

-> ListTagsForProjectResponse

Create a value of ListTagsForProjectResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:nextToken:ListTagsForProject', listTagsForProjectResponse_nextToken - Reserved for future use.

$sel:tags:ListTagsForProjectResponse', listTagsForProjectResponse_tags - The tags for the project.

$sel:httpStatus:ListTagsForProjectResponse', listTagsForProjectResponse_httpStatus - The response's http status code.

Response Lenses

listTagsForProjectResponse_nextToken :: Lens' ListTagsForProjectResponse (Maybe Text)

Reserved for future use.

listTagsForProjectResponse_tags :: Lens' ListTagsForProjectResponse (Maybe (HashMap Text Text))

The tags for the project.

listTagsForProjectResponse_httpStatus :: Lens' ListTagsForProjectResponse Int

The response's http status code.