Amazonka.CodeStar.ListTeamMembers

Description

Lists all team members associated with a project.

This operation returns paginated results.

Creating a Request

data ListTeamMembers

See: newListTeamMembers smart constructor.

Constructors

ListTeamMembers'

Fields

nextToken :: Maybe Text

The continuation token for the next set of results, if the results cannot be returned in one response.

maxResults :: Maybe Natural

The maximum number of team members you want returned in a response.

projectId :: Text

The ID of the project for which you want to list team members.

newListTeamMembers

Arguments

:: Text
$sel:projectId:ListTeamMembers'

-> ListTeamMembers

Create a value of ListTeamMembers with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:nextToken:ListTeamMembers', listTeamMembers_nextToken - The continuation token for the next set of results, if the results cannot be returned in one response.

$sel:maxResults:ListTeamMembers', listTeamMembers_maxResults - The maximum number of team members you want returned in a response.

$sel:projectId:ListTeamMembers', listTeamMembers_projectId - The ID of the project for which you want to list team members.

Request Lenses

listTeamMembers_nextToken :: Lens' ListTeamMembers (Maybe Text)

The continuation token for the next set of results, if the results cannot be returned in one response.

listTeamMembers_maxResults :: Lens' ListTeamMembers (Maybe Natural)

The maximum number of team members you want returned in a response.

listTeamMembers_projectId :: Lens' ListTeamMembers Text

The ID of the project for which you want to list team members.

Destructuring the Response

data ListTeamMembersResponse

See: newListTeamMembersResponse smart constructor.

Constructors

ListTeamMembersResponse'

Fields

nextToken :: Maybe Text

The continuation token to use when requesting the next set of results, if there are more results to be returned.

httpStatus :: Int

The response's http status code.

teamMembers :: [TeamMember]

A list of team member objects for the project.

newListTeamMembersResponse

Arguments

:: Int
$sel:httpStatus:ListTeamMembersResponse'

-> ListTeamMembersResponse

Create a value of ListTeamMembersResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:nextToken:ListTeamMembers', listTeamMembersResponse_nextToken - The continuation token to use when requesting the next set of results, if there are more results to be returned.

$sel:httpStatus:ListTeamMembersResponse', listTeamMembersResponse_httpStatus - The response's http status code.

$sel:teamMembers:ListTeamMembersResponse', listTeamMembersResponse_teamMembers - A list of team member objects for the project.

Response Lenses

listTeamMembersResponse_nextToken :: Lens' ListTeamMembersResponse (Maybe Text)

The continuation token to use when requesting the next set of results, if there are more results to be returned.

listTeamMembersResponse_httpStatus :: Lens' ListTeamMembersResponse Int

The response's http status code.

listTeamMembersResponse_teamMembers :: Lens' ListTeamMembersResponse [TeamMember]

A list of team member objects for the project.