Amazonka.CodeStar.Types.Code

Description

Documentation

data Code

Location and destination information about the source code files provided with the project request. The source code is uploaded to the new project source repository after project creation.

See: newCode smart constructor.

Constructors

Code'

Fields

source :: CodeSource

The location where the source code files provided with the project request are stored. AWS CodeStar retrieves the files during project creation.

destination :: CodeDestination

The repository to be created in AWS CodeStar. Valid values are AWS CodeCommit or GitHub. After AWS CodeStar provisions the new repository, the source code files provided with the project request are placed in the repository.

newCode

Arguments

:: CodeSource
$sel:source:Code'

-> CodeDestination
$sel:destination:Code'

-> Code

Create a value of Code with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:source:Code', code_source - The location where the source code files provided with the project request are stored. AWS CodeStar retrieves the files during project creation.

$sel:destination:Code', code_destination - The repository to be created in AWS CodeStar. Valid values are AWS CodeCommit or GitHub. After AWS CodeStar provisions the new repository, the source code files provided with the project request are placed in the repository.

code_source :: Lens' Code CodeSource

The location where the source code files provided with the project request are stored. AWS CodeStar retrieves the files during project creation.

code_destination :: Lens' Code CodeDestination

The repository to be created in AWS CodeStar. Valid values are AWS CodeCommit or GitHub. After AWS CodeStar provisions the new repository, the source code files provided with the project request are placed in the repository.