Amazonka.CodeStarConnections

Description

Derived from API version 2019-12-01 of the AWS service descriptions, licensed under Apache 2.0.

AWS CodeStar Connections

This AWS CodeStar Connections API Reference provides descriptions and usage examples of the operations and data types for the AWS CodeStar Connections API. You can use the connections API to work with connections and installations.

Connections are configurations that you use to connect AWS resources to external code repositories. Each connection is a resource that can be given to services such as CodePipeline to connect to a third-party repository such as Bitbucket. For example, you can add the connection in CodePipeline so that it triggers your pipeline when a code change is made to your third-party code repository. Each connection is named and associated with a unique ARN that is used to reference the connection.

When you create a connection, the console initiates a third-party connection handshake. Installations are the apps that are used to conduct this handshake. For example, the installation for the Bitbucket provider type is the Bitbucket app. When you create a connection, you can choose an existing installation or create one.

When you want to create a connection to an installed provider type such as GitHub Enterprise Server, you create a host for your connections.

You can work with connections by calling:

CreateConnection, which creates a uniquely named connection that can be referenced by services such as CodePipeline.

DeleteConnection, which deletes the specified connection.

GetConnection, which returns information about the connection, including the connection status.

ListConnections, which lists the connections associated with your account.

You can work with hosts by calling:

CreateHost, which creates a host that represents the infrastructure where your provider is installed.

DeleteHost, which deletes the specified host.

GetHost, which returns information about the host, including the setup status.

ListHosts, which lists the hosts associated with your account.

You can work with tags in AWS CodeStar Connections by calling the following:

ListTagsForResource, which gets information about AWS tags for a specified Amazon Resource Name (ARN) in AWS CodeStar Connections.

TagResource, which adds or updates tags for a resource in AWS CodeStar Connections.

UntagResource, which removes tags for a resource in AWS CodeStar Connections.

For information about how to use AWS CodeStar Connections, see the Developer Tools User Guide.

Service Configuration

defaultService :: Service

API version 2019-12-01 of the Amazon CodeStar connections SDK configuration.

Errors

Error matchers are designed for use with the functions provided by Control.Exception.Lens. This allows catching (and rethrowing) service specific errors returned by CodeStarConnections.

ResourceUnavailableException

_ResourceUnavailableException :: AsError a => Getting (First ServiceError) a ServiceError

Resource not found. Verify the ARN for the host resource and try again.

UnsupportedOperationException

_UnsupportedOperationException :: AsError a => Getting (First ServiceError) a ServiceError

The operation is not supported. Check the connection status and try again.

ConflictException

_ConflictException :: AsError a => Getting (First ServiceError) a ServiceError

Two conflicting operations have been made on the same resource.

ResourceNotFoundException

_ResourceNotFoundException :: AsError a => Getting (First ServiceError) a ServiceError

Resource not found. Verify the connection resource ARN and try again.

LimitExceededException

_LimitExceededException :: AsError a => Getting (First ServiceError) a ServiceError

Exceeded the maximum limit for connections.

Waiters

Waiters poll by repeatedly sending a request until some remote success condition configured by the Wait specification is fulfilled. The Wait specification determines how many attempts should be made, in addition to delay and retry strategies.

Operations

Some AWS operations return results that are incomplete and require subsequent requests in order to obtain the entire result set. The process of sending subsequent requests to continue where a previous request left off is called pagination. For example, the ListObjects operation of Amazon S3 returns up to 1000 objects at a time, and you must send subsequent requests with the appropriate Marker in order to retrieve the next page of results.

Operations that have an AWSPager instance can transparently perform subsequent requests, correctly setting Markers and other request facets to iterate through the entire result set of a truncated API operation. Operations which support this have an additional note in the documentation.

Many operations have the ability to filter results on the server side. See the individual operation parameters for details.

CreateHost

data CreateHost

See: newCreateHost smart constructor.

Constructors

CreateHost' (Maybe VpcConfiguration) (Maybe [Tag]) Text ProviderType Text

newCreateHost

Arguments

:: Text
$sel:name:CreateHost'

-> ProviderType
$sel:providerType:CreateHost'

-> Text
$sel:providerEndpoint:CreateHost'

-> CreateHost

Create a value of CreateHost with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:vpcConfiguration:CreateHost', createHost_vpcConfiguration - The VPC configuration to be provisioned for the host. A VPC must be configured and the infrastructure to be represented by the host must already be connected to the VPC.

$sel:tags:CreateHost', createHost_tags - Undocumented member.

$sel:name:CreateHost', createHost_name - The name of the host to be created. The name must be unique in the calling AWS account.

$sel:providerType:CreateHost', createHost_providerType - The name of the installed provider to be associated with your connection. The host resource represents the infrastructure where your provider type is installed. The valid provider type is GitHub Enterprise Server.

$sel:providerEndpoint:CreateHost', createHost_providerEndpoint - The endpoint of the infrastructure to be represented by the host after it is created.

data CreateHostResponse

See: newCreateHostResponse smart constructor.

Constructors

CreateHostResponse' (Maybe Text) (Maybe [Tag]) Int

newCreateHostResponse

Arguments

:: Int
$sel:httpStatus:CreateHostResponse'

-> CreateHostResponse

Create a value of CreateHostResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:hostArn:CreateHostResponse', createHostResponse_hostArn - The Amazon Resource Name (ARN) of the host to be created.

$sel:tags:CreateHost', createHostResponse_tags - Undocumented member.

$sel:httpStatus:CreateHostResponse', createHostResponse_httpStatus - The response's http status code.

ListConnections

data ListConnections

See: newListConnections smart constructor.

Constructors

ListConnections' (Maybe Text) (Maybe Text) (Maybe Natural) (Maybe ProviderType)

newListConnections :: ListConnections

Create a value of ListConnections with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:hostArnFilter:ListConnections', listConnections_hostArnFilter - Filters the list of connections to those associated with a specified host.

$sel:nextToken:ListConnections', listConnections_nextToken - The token that was returned from the previous ListConnections call, which can be used to return the next set of connections in the list.

$sel:maxResults:ListConnections', listConnections_maxResults - The maximum number of results to return in a single call. To retrieve the remaining results, make another call with the returned nextToken value.

$sel:providerTypeFilter:ListConnections', listConnections_providerTypeFilter - Filters the list of connections to those associated with a specified provider, such as Bitbucket.

data ListConnectionsResponse

See: newListConnectionsResponse smart constructor.

Constructors

ListConnectionsResponse' (Maybe [Connection]) (Maybe Text) Int

newListConnectionsResponse

Arguments

:: Int
$sel:httpStatus:ListConnectionsResponse'

-> ListConnectionsResponse

Create a value of ListConnectionsResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:connections:ListConnectionsResponse', listConnectionsResponse_connections - A list of connections and the details for each connection, such as status, owner, and provider type.

$sel:nextToken:ListConnections', listConnectionsResponse_nextToken - A token that can be used in the next ListConnections call. To view all items in the list, continue to call this operation with each subsequent token until no more nextToken values are returned.

$sel:httpStatus:ListConnectionsResponse', listConnectionsResponse_httpStatus - The response's http status code.

DeleteConnection

data DeleteConnection

See: newDeleteConnection smart constructor.

Constructors

DeleteConnection' Text

newDeleteConnection

Arguments

:: Text
$sel:connectionArn:DeleteConnection'

-> DeleteConnection

Create a value of DeleteConnection with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:connectionArn:DeleteConnection', deleteConnection_connectionArn - The Amazon Resource Name (ARN) of the connection to be deleted.

The ARN is never reused if the connection is deleted.

data DeleteConnectionResponse

See: newDeleteConnectionResponse smart constructor.

Constructors

DeleteConnectionResponse' Int

newDeleteConnectionResponse

Arguments

:: Int
$sel:httpStatus:DeleteConnectionResponse'

-> DeleteConnectionResponse

Create a value of DeleteConnectionResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:httpStatus:DeleteConnectionResponse', deleteConnectionResponse_httpStatus - The response's http status code.

ListTagsForResource

data ListTagsForResource

See: newListTagsForResource smart constructor.

Constructors

ListTagsForResource' Text

newListTagsForResource

Arguments

:: Text
$sel:resourceArn:ListTagsForResource'

-> ListTagsForResource

Create a value of ListTagsForResource with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:resourceArn:ListTagsForResource', listTagsForResource_resourceArn - The Amazon Resource Name (ARN) of the resource for which you want to get information about tags, if any.

data ListTagsForResourceResponse

See: newListTagsForResourceResponse smart constructor.

Constructors

ListTagsForResourceResponse' (Maybe [Tag]) Int

newListTagsForResourceResponse

Arguments

:: Int
$sel:httpStatus:ListTagsForResourceResponse'

-> ListTagsForResourceResponse

Create a value of ListTagsForResourceResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:tags:ListTagsForResourceResponse', listTagsForResourceResponse_tags - A list of tag key and value pairs associated with the specified resource.

$sel:httpStatus:ListTagsForResourceResponse', listTagsForResourceResponse_httpStatus - The response's http status code.

CreateConnection

data CreateConnection

See: newCreateConnection smart constructor.

Constructors

CreateConnection' (Maybe ProviderType) (Maybe Text) (Maybe [Tag]) Text

newCreateConnection

Arguments

:: Text
$sel:connectionName:CreateConnection'

-> CreateConnection

Create a value of CreateConnection with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:providerType:CreateConnection', createConnection_providerType - The name of the external provider where your third-party code repository is configured.

$sel:hostArn:CreateConnection', createConnection_hostArn - The Amazon Resource Name (ARN) of the host associated with the connection to be created.

$sel:tags:CreateConnection', createConnection_tags - The key-value pair to use when tagging the resource.

$sel:connectionName:CreateConnection', createConnection_connectionName - The name of the connection to be created. The name must be unique in the calling AWS account.

data CreateConnectionResponse

See: newCreateConnectionResponse smart constructor.

Constructors

CreateConnectionResponse' (Maybe [Tag]) Int Text

newCreateConnectionResponse

Arguments

:: Int
$sel:httpStatus:CreateConnectionResponse'

-> Text
$sel:connectionArn:CreateConnectionResponse'

-> CreateConnectionResponse

Create a value of CreateConnectionResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:tags:CreateConnection', createConnectionResponse_tags - Specifies the tags applied to the resource.

$sel:httpStatus:CreateConnectionResponse', createConnectionResponse_httpStatus - The response's http status code.

$sel:connectionArn:CreateConnectionResponse', createConnectionResponse_connectionArn - The Amazon Resource Name (ARN) of the connection to be created. The ARN is used as the connection reference when the connection is shared between AWS services.

The ARN is never reused if the connection is deleted.

GetConnection

data GetConnection

See: newGetConnection smart constructor.

Constructors

GetConnection' Text

newGetConnection

Arguments

:: Text
$sel:connectionArn:GetConnection'

-> GetConnection

Create a value of GetConnection with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:connectionArn:GetConnection', getConnection_connectionArn - The Amazon Resource Name (ARN) of a connection.

data GetConnectionResponse

See: newGetConnectionResponse smart constructor.

Constructors

GetConnectionResponse' (Maybe Connection) Int

newGetConnectionResponse

Arguments

:: Int
$sel:httpStatus:GetConnectionResponse'

-> GetConnectionResponse

Create a value of GetConnectionResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:connection:GetConnectionResponse', getConnectionResponse_connection - The connection details, such as status, owner, and provider type.

$sel:httpStatus:GetConnectionResponse', getConnectionResponse_httpStatus - The response's http status code.

DeleteHost

data DeleteHost

See: newDeleteHost smart constructor.

Constructors

DeleteHost' Text

newDeleteHost

Arguments

:: Text
$sel:hostArn:DeleteHost'

-> DeleteHost

Create a value of DeleteHost with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:hostArn:DeleteHost', deleteHost_hostArn - The Amazon Resource Name (ARN) of the host to be deleted.

data DeleteHostResponse

See: newDeleteHostResponse smart constructor.

Constructors

DeleteHostResponse' Int

newDeleteHostResponse

Arguments

:: Int
$sel:httpStatus:DeleteHostResponse'

-> DeleteHostResponse

Create a value of DeleteHostResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:httpStatus:DeleteHostResponse', deleteHostResponse_httpStatus - The response's http status code.

UpdateHost

data UpdateHost

See: newUpdateHost smart constructor.

Constructors

UpdateHost' (Maybe Text) (Maybe VpcConfiguration) Text

newUpdateHost

Arguments

:: Text
$sel:hostArn:UpdateHost'

-> UpdateHost

Create a value of UpdateHost with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:providerEndpoint:UpdateHost', updateHost_providerEndpoint - The URL or endpoint of the host to be updated.

$sel:vpcConfiguration:UpdateHost', updateHost_vpcConfiguration - The VPC configuration of the host to be updated. A VPC must be configured and the infrastructure to be represented by the host must already be connected to the VPC.

$sel:hostArn:UpdateHost', updateHost_hostArn - The Amazon Resource Name (ARN) of the host to be updated.

data UpdateHostResponse

See: newUpdateHostResponse smart constructor.

Constructors

UpdateHostResponse' Int

newUpdateHostResponse

Arguments

:: Int
$sel:httpStatus:UpdateHostResponse'

-> UpdateHostResponse

Create a value of UpdateHostResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:httpStatus:UpdateHostResponse', updateHostResponse_httpStatus - The response's http status code.

ListHosts

data ListHosts

See: newListHosts smart constructor.

Constructors

ListHosts' (Maybe Text) (Maybe Natural)

newListHosts :: ListHosts

Create a value of ListHosts with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:nextToken:ListHosts', listHosts_nextToken - The token that was returned from the previous ListHosts call, which can be used to return the next set of hosts in the list.

$sel:maxResults:ListHosts', listHosts_maxResults - The maximum number of results to return in a single call. To retrieve the remaining results, make another call with the returned nextToken value.

data ListHostsResponse

See: newListHostsResponse smart constructor.

Constructors

ListHostsResponse' (Maybe [Host]) (Maybe Text) Int

newListHostsResponse

Arguments

:: Int
$sel:httpStatus:ListHostsResponse'

-> ListHostsResponse

Create a value of ListHostsResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:hosts:ListHostsResponse', listHostsResponse_hosts - A list of hosts and the details for each host, such as status, endpoint, and provider type.

$sel:nextToken:ListHosts', listHostsResponse_nextToken - A token that can be used in the next ListHosts call. To view all items in the list, continue to call this operation with each subsequent token until no more nextToken values are returned.

$sel:httpStatus:ListHostsResponse', listHostsResponse_httpStatus - The response's http status code.

TagResource

data TagResource

See: newTagResource smart constructor.

Constructors

TagResource' Text [Tag]

newTagResource

Arguments

:: Text
$sel:resourceArn:TagResource'

-> TagResource

Create a value of TagResource with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:resourceArn:TagResource', tagResource_resourceArn - The Amazon Resource Name (ARN) of the resource to which you want to add or update tags.

$sel:tags:TagResource', tagResource_tags - The tags you want to modify or add to the resource.

data TagResourceResponse

See: newTagResourceResponse smart constructor.

Constructors

TagResourceResponse' Int

newTagResourceResponse

Arguments

:: Int
$sel:httpStatus:TagResourceResponse'

-> TagResourceResponse

Create a value of TagResourceResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:httpStatus:TagResourceResponse', tagResourceResponse_httpStatus - The response's http status code.

UntagResource

data UntagResource

See: newUntagResource smart constructor.

Constructors

UntagResource' Text [Text]

newUntagResource

Arguments

:: Text
$sel:resourceArn:UntagResource'

-> UntagResource

Create a value of UntagResource with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:resourceArn:UntagResource', untagResource_resourceArn - The Amazon Resource Name (ARN) of the resource to remove tags from.

$sel:tagKeys:UntagResource', untagResource_tagKeys - The list of keys for the tags to be removed from the resource.

data UntagResourceResponse

See: newUntagResourceResponse smart constructor.

Constructors

UntagResourceResponse' Int

newUntagResourceResponse

Arguments

:: Int
$sel:httpStatus:UntagResourceResponse'

-> UntagResourceResponse

Create a value of UntagResourceResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:httpStatus:UntagResourceResponse', untagResourceResponse_httpStatus - The response's http status code.

GetHost

data GetHost

See: newGetHost smart constructor.

Constructors

GetHost' Text

newGetHost

Arguments

:: Text
$sel:hostArn:GetHost'

-> GetHost

Create a value of GetHost with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:hostArn:GetHost', getHost_hostArn - The Amazon Resource Name (ARN) of the requested host.

data GetHostResponse

See: newGetHostResponse smart constructor.

Constructors

GetHostResponse' (Maybe Text) (Maybe Text) (Maybe ProviderType) (Maybe Text) (Maybe VpcConfiguration) Int

newGetHostResponse

Arguments

:: Int
$sel:httpStatus:GetHostResponse'

-> GetHostResponse

Create a value of GetHostResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:status:GetHostResponse', getHostResponse_status - The status of the requested host.

$sel:providerEndpoint:GetHostResponse', getHostResponse_providerEndpoint - The endpoint of the infrastructure represented by the requested host.

$sel:providerType:GetHostResponse', getHostResponse_providerType - The provider type of the requested host, such as GitHub Enterprise Server.

$sel:name:GetHostResponse', getHostResponse_name - The name of the requested host.

$sel:vpcConfiguration:GetHostResponse', getHostResponse_vpcConfiguration - The VPC configuration of the requested host.

$sel:httpStatus:GetHostResponse', getHostResponse_httpStatus - The response's http status code.

Types

ConnectionStatus

newtype ConnectionStatus

Constructors

ConnectionStatus'

Fields

fromConnectionStatus :: Text

Bundled Patterns

pattern ConnectionStatus_AVAILABLE :: ConnectionStatus

pattern ConnectionStatus_ERROR :: ConnectionStatus

pattern ConnectionStatus_PENDING :: ConnectionStatus

ProviderType

newtype ProviderType

Constructors

ProviderType'

Fields

fromProviderType :: Text

Bundled Patterns

pattern ProviderType_Bitbucket :: ProviderType

pattern ProviderType_GitHub :: ProviderType

pattern ProviderType_GitHubEnterpriseServer :: ProviderType

Connection

data Connection

A resource that is used to connect third-party source providers with services like AWS CodePipeline.

Note: A connection created through CloudFormation, the CLI, or the SDK is in `PENDING` status by default. You can make its status `AVAILABLE` by updating the connection in the console.

See: newConnection smart constructor.

Constructors

Connection' (Maybe Text) (Maybe ProviderType) (Maybe Text) (Maybe ConnectionStatus) (Maybe Text) (Maybe Text)

newConnection :: Connection

Create a value of Connection with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:ownerAccountId:Connection', connection_ownerAccountId - The identifier of the external provider where your third-party code repository is configured. For Bitbucket, this is the account ID of the owner of the Bitbucket repository.

$sel:providerType:Connection', connection_providerType - The name of the external provider where your third-party code repository is configured.

$sel:connectionName:Connection', connection_connectionName - The name of the connection. Connection names must be unique in an AWS user account.

$sel:connectionStatus:Connection', connection_connectionStatus - The current status of the connection.

$sel:hostArn:Connection', connection_hostArn - The Amazon Resource Name (ARN) of the host associated with the connection.

$sel:connectionArn:Connection', connection_connectionArn - The Amazon Resource Name (ARN) of the connection. The ARN is used as the connection reference when the connection is shared between AWS services.

The ARN is never reused if the connection is deleted.

Host

data Host

A resource that represents the infrastructure where a third-party provider is installed. The host is used when you create connections to an installed third-party provider type, such as GitHub Enterprise Server. You create one host for all connections to that provider.

A host created through the CLI or the SDK is in `PENDING` status by default. You can make its status `AVAILABLE` by setting up the host in the console.

See: newHost smart constructor.

Constructors

Host' (Maybe Text) (Maybe Text) (Maybe ProviderType) (Maybe Text) (Maybe Text) (Maybe Text) (Maybe VpcConfiguration)

newHost :: Host

Create a value of Host with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:status:Host', host_status - The status of the host, such as PENDING, AVAILABLE, VPC_CONFIG_DELETING, VPC_CONFIG_INITIALIZING, and VPC_CONFIG_FAILED_INITIALIZATION.

$sel:providerEndpoint:Host', host_providerEndpoint - The endpoint of the infrastructure where your provider type is installed.

$sel:providerType:Host', host_providerType - The name of the installed provider to be associated with your connection. The host resource represents the infrastructure where your provider type is installed. The valid provider type is GitHub Enterprise Server.

$sel:statusMessage:Host', host_statusMessage - The status description for the host.

$sel:name:Host', host_name - The name of the host.

$sel:hostArn:Host', host_hostArn - The Amazon Resource Name (ARN) of the host.

$sel:vpcConfiguration:Host', host_vpcConfiguration - The VPC configuration provisioned for the host.

Tag

data Tag

A tag is a key-value pair that is used to manage the resource.

This tag is available for use by AWS services that support tags.

See: newTag smart constructor.

Constructors

Tag' Text Text

newTag

Arguments

:: Text
$sel:key:Tag'

-> Text
$sel:value:Tag'

-> Tag

Create a value of Tag with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:key:Tag', tag_key - The tag's key.

$sel:value:Tag', tag_value - The tag's value.

VpcConfiguration

data VpcConfiguration

The VPC configuration provisioned for the host.

See: newVpcConfiguration smart constructor.

Constructors

VpcConfiguration' (Maybe Text) Text (NonEmpty Text) (NonEmpty Text)

newVpcConfiguration

Arguments

:: Text
$sel:vpcId:VpcConfiguration'

-> NonEmpty Text
$sel:subnetIds:VpcConfiguration'

-> NonEmpty Text
$sel:securityGroupIds:VpcConfiguration'

-> VpcConfiguration

Create a value of VpcConfiguration with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:tlsCertificate:VpcConfiguration', vpcConfiguration_tlsCertificate - The value of the Transport Layer Security (TLS) certificate associated with the infrastructure where your provider type is installed.

$sel:vpcId:VpcConfiguration', vpcConfiguration_vpcId - The ID of the Amazon VPC connected to the infrastructure where your provider type is installed.

$sel:subnetIds:VpcConfiguration', vpcConfiguration_subnetIds - The ID of the subnet or subnets associated with the Amazon VPC connected to the infrastructure where your provider type is installed.

$sel:securityGroupIds:VpcConfiguration', vpcConfiguration_securityGroupIds - The ID of the security group or security groups associated with the Amazon VPC connected to the infrastructure where your provider type is installed.