Amazonka.CodeStarConnections.CreateHost

Description

Creates a resource that represents the infrastructure where a third-party provider is installed. The host is used when you create connections to an installed third-party provider type, such as GitHub Enterprise Server. You create one host for all connections to that provider.

A host created through the CLI or the SDK is in `PENDING` status by default. You can make its status `AVAILABLE` by setting up the host in the console.

Creating a Request

data CreateHost

See: newCreateHost smart constructor.

Constructors

CreateHost'

Fields

vpcConfiguration :: Maybe VpcConfiguration

The VPC configuration to be provisioned for the host. A VPC must be configured and the infrastructure to be represented by the host must already be connected to the VPC.

tags :: Maybe [Tag]

name :: Text

The name of the host to be created. The name must be unique in the calling AWS account.

providerType :: ProviderType

The name of the installed provider to be associated with your connection. The host resource represents the infrastructure where your provider type is installed. The valid provider type is GitHub Enterprise Server.

providerEndpoint :: Text

The endpoint of the infrastructure to be represented by the host after it is created.

newCreateHost

Arguments

:: Text
$sel:name:CreateHost'

-> ProviderType
$sel:providerType:CreateHost'

-> Text
$sel:providerEndpoint:CreateHost'

-> CreateHost

Create a value of CreateHost with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:vpcConfiguration:CreateHost', createHost_vpcConfiguration - The VPC configuration to be provisioned for the host. A VPC must be configured and the infrastructure to be represented by the host must already be connected to the VPC.

$sel:tags:CreateHost', createHost_tags - Undocumented member.

$sel:name:CreateHost', createHost_name - The name of the host to be created. The name must be unique in the calling AWS account.

$sel:providerType:CreateHost', createHost_providerType - The name of the installed provider to be associated with your connection. The host resource represents the infrastructure where your provider type is installed. The valid provider type is GitHub Enterprise Server.

$sel:providerEndpoint:CreateHost', createHost_providerEndpoint - The endpoint of the infrastructure to be represented by the host after it is created.

Request Lenses

createHost_vpcConfiguration :: Lens' CreateHost (Maybe VpcConfiguration)

The VPC configuration to be provisioned for the host. A VPC must be configured and the infrastructure to be represented by the host must already be connected to the VPC.

createHost_tags :: Lens' CreateHost (Maybe [Tag])

Undocumented member.

createHost_name :: Lens' CreateHost Text

The name of the host to be created. The name must be unique in the calling AWS account.

createHost_providerType :: Lens' CreateHost ProviderType

The name of the installed provider to be associated with your connection. The host resource represents the infrastructure where your provider type is installed. The valid provider type is GitHub Enterprise Server.

createHost_providerEndpoint :: Lens' CreateHost Text

The endpoint of the infrastructure to be represented by the host after it is created.

Destructuring the Response

data CreateHostResponse

See: newCreateHostResponse smart constructor.

Constructors

CreateHostResponse'

Fields

hostArn :: Maybe Text

The Amazon Resource Name (ARN) of the host to be created.

tags :: Maybe [Tag]

httpStatus :: Int

The response's http status code.

newCreateHostResponse

Arguments

:: Int
$sel:httpStatus:CreateHostResponse'

-> CreateHostResponse

Create a value of CreateHostResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:hostArn:CreateHostResponse', createHostResponse_hostArn - The Amazon Resource Name (ARN) of the host to be created.

$sel:tags:CreateHost', createHostResponse_tags - Undocumented member.

$sel:httpStatus:CreateHostResponse', createHostResponse_httpStatus - The response's http status code.

Response Lenses

createHostResponse_hostArn :: Lens' CreateHostResponse (Maybe Text)

The Amazon Resource Name (ARN) of the host to be created.

createHostResponse_tags :: Lens' CreateHostResponse (Maybe [Tag])

Undocumented member.

createHostResponse_httpStatus :: Lens' CreateHostResponse Int

The response's http status code.