Amazonka.CodeStarConnections.ListHosts

Description

Lists the hosts associated with your account.

Creating a Request

data ListHosts

See: newListHosts smart constructor.

Constructors

ListHosts'

Fields

nextToken :: Maybe Text

The token that was returned from the previous ListHosts call, which can be used to return the next set of hosts in the list.

maxResults :: Maybe Natural

The maximum number of results to return in a single call. To retrieve the remaining results, make another call with the returned nextToken value.

newListHosts :: ListHosts

Create a value of ListHosts with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:nextToken:ListHosts', listHosts_nextToken - The token that was returned from the previous ListHosts call, which can be used to return the next set of hosts in the list.

$sel:maxResults:ListHosts', listHosts_maxResults - The maximum number of results to return in a single call. To retrieve the remaining results, make another call with the returned nextToken value.

Request Lenses

listHosts_nextToken :: Lens' ListHosts (Maybe Text)

The token that was returned from the previous ListHosts call, which can be used to return the next set of hosts in the list.

listHosts_maxResults :: Lens' ListHosts (Maybe Natural)

The maximum number of results to return in a single call. To retrieve the remaining results, make another call with the returned nextToken value.

Destructuring the Response

data ListHostsResponse

See: newListHostsResponse smart constructor.

Constructors

ListHostsResponse'

Fields

hosts :: Maybe [Host]

A list of hosts and the details for each host, such as status, endpoint, and provider type.

nextToken :: Maybe Text

A token that can be used in the next ListHosts call. To view all items in the list, continue to call this operation with each subsequent token until no more nextToken values are returned.

httpStatus :: Int

The response's http status code.

newListHostsResponse

Arguments

:: Int
$sel:httpStatus:ListHostsResponse'

-> ListHostsResponse

Create a value of ListHostsResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:hosts:ListHostsResponse', listHostsResponse_hosts - A list of hosts and the details for each host, such as status, endpoint, and provider type.

$sel:nextToken:ListHosts', listHostsResponse_nextToken - A token that can be used in the next ListHosts call. To view all items in the list, continue to call this operation with each subsequent token until no more nextToken values are returned.

$sel:httpStatus:ListHostsResponse', listHostsResponse_httpStatus - The response's http status code.

Response Lenses

listHostsResponse_hosts :: Lens' ListHostsResponse (Maybe [Host])

A list of hosts and the details for each host, such as status, endpoint, and provider type.

listHostsResponse_nextToken :: Lens' ListHostsResponse (Maybe Text)

A token that can be used in the next ListHosts call. To view all items in the list, continue to call this operation with each subsequent token until no more nextToken values are returned.

listHostsResponse_httpStatus :: Lens' ListHostsResponse Int

The response's http status code.