Amazonka.CodeStarConnections.Types.Connection

Description

Documentation

data Connection

A resource that is used to connect third-party source providers with services like AWS CodePipeline.

Note: A connection created through CloudFormation, the CLI, or the SDK is in `PENDING` status by default. You can make its status `AVAILABLE` by updating the connection in the console.

See: newConnection smart constructor.

Constructors

Connection'

Fields

ownerAccountId :: Maybe Text

The identifier of the external provider where your third-party code repository is configured. For Bitbucket, this is the account ID of the owner of the Bitbucket repository.

providerType :: Maybe ProviderType

The name of the external provider where your third-party code repository is configured.

connectionName :: Maybe Text

The name of the connection. Connection names must be unique in an AWS user account.

connectionStatus :: Maybe ConnectionStatus

The current status of the connection.

hostArn :: Maybe Text

The Amazon Resource Name (ARN) of the host associated with the connection.

connectionArn :: Maybe Text

The Amazon Resource Name (ARN) of the connection. The ARN is used as the connection reference when the connection is shared between AWS services.

The ARN is never reused if the connection is deleted.

newConnection :: Connection

Create a value of Connection with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:ownerAccountId:Connection', connection_ownerAccountId - The identifier of the external provider where your third-party code repository is configured. For Bitbucket, this is the account ID of the owner of the Bitbucket repository.

$sel:providerType:Connection', connection_providerType - The name of the external provider where your third-party code repository is configured.

$sel:connectionName:Connection', connection_connectionName - The name of the connection. Connection names must be unique in an AWS user account.

$sel:connectionStatus:Connection', connection_connectionStatus - The current status of the connection.

$sel:hostArn:Connection', connection_hostArn - The Amazon Resource Name (ARN) of the host associated with the connection.

$sel:connectionArn:Connection', connection_connectionArn - The Amazon Resource Name (ARN) of the connection. The ARN is used as the connection reference when the connection is shared between AWS services.

The ARN is never reused if the connection is deleted.

connection_ownerAccountId :: Lens' Connection (Maybe Text)

The identifier of the external provider where your third-party code repository is configured. For Bitbucket, this is the account ID of the owner of the Bitbucket repository.

connection_providerType :: Lens' Connection (Maybe ProviderType)

The name of the external provider where your third-party code repository is configured.

connection_connectionName :: Lens' Connection (Maybe Text)

The name of the connection. Connection names must be unique in an AWS user account.

connection_connectionStatus :: Lens' Connection (Maybe ConnectionStatus)

The current status of the connection.

connection_hostArn :: Lens' Connection (Maybe Text)

The Amazon Resource Name (ARN) of the host associated with the connection.

connection_connectionArn :: Lens' Connection (Maybe Text)

The Amazon Resource Name (ARN) of the connection. The ARN is used as the connection reference when the connection is shared between AWS services.

The ARN is never reused if the connection is deleted.