Amazonka.CodeStarNotifications.Lens

Description

Operations

CreateNotificationRule

createNotificationRule_status :: Lens' CreateNotificationRule (Maybe NotificationRuleStatus)

The status of the notification rule. The default value is ENABLED. If the status is set to DISABLED, notifications aren't sent for the notification rule.

createNotificationRule_clientRequestToken :: Lens' CreateNotificationRule (Maybe Text)

A unique, client-generated idempotency token that, when provided in a request, ensures the request cannot be repeated with a changed parameter. If a request with the same parameters is received and a token is included, the request returns information about the initial request that used that token.

The AWS SDKs prepopulate client request tokens. If you are using an AWS SDK, an idempotency token is created for you.

createNotificationRule_tags :: Lens' CreateNotificationRule (Maybe (HashMap Text Text))

A list of tags to apply to this notification rule. Key names cannot start with "aws".

createNotificationRule_name :: Lens' CreateNotificationRule Text

The name for the notification rule. Notifictaion rule names must be unique in your AWS account.

createNotificationRule_eventTypeIds :: Lens' CreateNotificationRule [Text]

A list of event types associated with this notification rule. For a list of allowed events, see EventTypeSummary.

createNotificationRule_resource :: Lens' CreateNotificationRule Text

The Amazon Resource Name (ARN) of the resource to associate with the notification rule. Supported resources include pipelines in AWS CodePipeline, repositories in AWS CodeCommit, and build projects in AWS CodeBuild.

createNotificationRule_targets :: Lens' CreateNotificationRule [Target]

A list of Amazon Resource Names (ARNs) of SNS topics to associate with the notification rule.

createNotificationRule_detailType :: Lens' CreateNotificationRule DetailType

The level of detail to include in the notifications for this resource. BASIC will include only the contents of the event as it would appear in AWS CloudWatch. FULL will include any supplemental information provided by AWS CodeStar Notifications and/or the service for the resource for which the notification is created.

createNotificationRuleResponse_arn :: Lens' CreateNotificationRuleResponse (Maybe Text)

The Amazon Resource Name (ARN) of the notification rule.

createNotificationRuleResponse_httpStatus :: Lens' CreateNotificationRuleResponse Int

The response's http status code.

UpdateNotificationRule

updateNotificationRule_status :: Lens' UpdateNotificationRule (Maybe NotificationRuleStatus)

The status of the notification rule. Valid statuses include enabled (sending notifications) or disabled (not sending notifications).

updateNotificationRule_eventTypeIds :: Lens' UpdateNotificationRule (Maybe [Text])

A list of event types associated with this notification rule.

updateNotificationRule_detailType :: Lens' UpdateNotificationRule (Maybe DetailType)

The level of detail to include in the notifications for this resource. BASIC will include only the contents of the event as it would appear in AWS CloudWatch. FULL will include any supplemental information provided by AWS CodeStar Notifications and/or the service for the resource for which the notification is created.

updateNotificationRule_name :: Lens' UpdateNotificationRule (Maybe Text)

The name of the notification rule.

updateNotificationRule_targets :: Lens' UpdateNotificationRule (Maybe [Target])

The address and type of the targets to receive notifications from this notification rule.

updateNotificationRule_arn :: Lens' UpdateNotificationRule Text

The Amazon Resource Name (ARN) of the notification rule.

updateNotificationRuleResponse_httpStatus :: Lens' UpdateNotificationRuleResponse Int

The response's http status code.

DeleteNotificationRule

deleteNotificationRule_arn :: Lens' DeleteNotificationRule Text

The Amazon Resource Name (ARN) of the notification rule you want to delete.

deleteNotificationRuleResponse_arn :: Lens' DeleteNotificationRuleResponse (Maybe Text)

The Amazon Resource Name (ARN) of the deleted notification rule.

deleteNotificationRuleResponse_httpStatus :: Lens' DeleteNotificationRuleResponse Int

The response's http status code.

ListTagsForResource

listTagsForResource_arn :: Lens' ListTagsForResource Text

The Amazon Resource Name (ARN) for the notification rule.

listTagsForResourceResponse_tags :: Lens' ListTagsForResourceResponse (Maybe (HashMap Text Text))

The tags associated with the notification rule.

listTagsForResourceResponse_httpStatus :: Lens' ListTagsForResourceResponse Int

The response's http status code.

ListEventTypes

listEventTypes_filters :: Lens' ListEventTypes (Maybe [ListEventTypesFilter])

The filters to use to return information by service or resource type.

listEventTypes_nextToken :: Lens' ListEventTypes (Maybe Text)

An enumeration token that, when provided in a request, returns the next batch of the results.

listEventTypes_maxResults :: Lens' ListEventTypes (Maybe Natural)

A non-negative integer used to limit the number of returned results. The default number is 50. The maximum number of results that can be returned is 100.

listEventTypesResponse_eventTypes :: Lens' ListEventTypesResponse (Maybe [EventTypeSummary])

Information about each event, including service name, resource type, event ID, and event name.

listEventTypesResponse_nextToken :: Lens' ListEventTypesResponse (Maybe Text)

An enumeration token that can be used in a request to return the next batch of the results.

listEventTypesResponse_httpStatus :: Lens' ListEventTypesResponse Int

The response's http status code.

DeleteTarget

deleteTarget_forceUnsubscribeAll :: Lens' DeleteTarget (Maybe Bool)

A Boolean value that can be used to delete all associations with this SNS topic. The default value is FALSE. If set to TRUE, all associations between that target and every notification rule in your AWS account are deleted.

deleteTarget_targetAddress :: Lens' DeleteTarget Text

The Amazon Resource Name (ARN) of the SNS topic to delete.

deleteTargetResponse_httpStatus :: Lens' DeleteTargetResponse Int

The response's http status code.

ListNotificationRules

listNotificationRules_filters :: Lens' ListNotificationRules (Maybe [ListNotificationRulesFilter])

The filters to use to return information by service or resource type. For valid values, see ListNotificationRulesFilter.

A filter with the same name can appear more than once when used with OR statements. Filters with different names should be applied with AND statements.

listNotificationRules_nextToken :: Lens' ListNotificationRules (Maybe Text)

An enumeration token that, when provided in a request, returns the next batch of the results.

listNotificationRules_maxResults :: Lens' ListNotificationRules (Maybe Natural)

A non-negative integer used to limit the number of returned results. The maximum number of results that can be returned is 100.

listNotificationRulesResponse_nextToken :: Lens' ListNotificationRulesResponse (Maybe Text)

An enumeration token that can be used in a request to return the next batch of the results.

listNotificationRulesResponse_notificationRules :: Lens' ListNotificationRulesResponse (Maybe [NotificationRuleSummary])

The list of notification rules for the AWS account, by Amazon Resource Name (ARN) and ID.

listNotificationRulesResponse_httpStatus :: Lens' ListNotificationRulesResponse Int

The response's http status code.

ListTargets

listTargets_filters :: Lens' ListTargets (Maybe [ListTargetsFilter])

The filters to use to return information by service or resource type. Valid filters include target type, target address, and target status.

A filter with the same name can appear more than once when used with OR statements. Filters with different names should be applied with AND statements.

listTargets_nextToken :: Lens' ListTargets (Maybe Text)

An enumeration token that, when provided in a request, returns the next batch of the results.

listTargets_maxResults :: Lens' ListTargets (Maybe Natural)

A non-negative integer used to limit the number of returned results. The maximum number of results that can be returned is 100.

listTargetsResponse_nextToken :: Lens' ListTargetsResponse (Maybe Text)

An enumeration token that can be used in a request to return the next batch of results.

listTargetsResponse_targets :: Lens' ListTargetsResponse (Maybe [TargetSummary])

The list of notification rule targets.

listTargetsResponse_httpStatus :: Lens' ListTargetsResponse Int

The response's http status code.

TagResource

tagResource_arn :: Lens' TagResource Text

The Amazon Resource Name (ARN) of the notification rule to tag.

tagResource_tags :: Lens' TagResource (HashMap Text Text)

The list of tags to associate with the resource. Tag key names cannot start with "aws".

tagResourceResponse_tags :: Lens' TagResourceResponse (Maybe (HashMap Text Text))

The list of tags associated with the resource.

tagResourceResponse_httpStatus :: Lens' TagResourceResponse Int

The response's http status code.

Subscribe

subscribe_clientRequestToken :: Lens' Subscribe (Maybe Text)

An enumeration token that, when provided in a request, returns the next batch of the results.

subscribe_arn :: Lens' Subscribe Text

The Amazon Resource Name (ARN) of the notification rule for which you want to create the association.

subscribe_target :: Lens' Subscribe Target

Undocumented member.

subscribeResponse_arn :: Lens' SubscribeResponse (Maybe Text)

The Amazon Resource Name (ARN) of the notification rule for which you have created assocations.

subscribeResponse_httpStatus :: Lens' SubscribeResponse Int

The response's http status code.

UntagResource

untagResource_arn :: Lens' UntagResource Text

The Amazon Resource Name (ARN) of the notification rule from which to remove the tags.

untagResource_tagKeys :: Lens' UntagResource [Text]

The key names of the tags to remove.

untagResourceResponse_httpStatus :: Lens' UntagResourceResponse Int

The response's http status code.

Unsubscribe

unsubscribe_arn :: Lens' Unsubscribe Text

The Amazon Resource Name (ARN) of the notification rule.

unsubscribe_targetAddress :: Lens' Unsubscribe Text

The ARN of the SNS topic to unsubscribe from the notification rule.

unsubscribeResponse_httpStatus :: Lens' UnsubscribeResponse Int

The response's http status code.

unsubscribeResponse_arn :: Lens' UnsubscribeResponse Text

The Amazon Resource Name (ARN) of the the notification rule from which you have removed a subscription.

DescribeNotificationRule

describeNotificationRule_arn :: Lens' DescribeNotificationRule Text

The Amazon Resource Name (ARN) of the notification rule.

describeNotificationRuleResponse_status :: Lens' DescribeNotificationRuleResponse (Maybe NotificationRuleStatus)

The status of the notification rule. Valid statuses are on (sending notifications) or off (not sending notifications).

describeNotificationRuleResponse_eventTypes :: Lens' DescribeNotificationRuleResponse (Maybe [EventTypeSummary])

A list of the event types associated with the notification rule.

describeNotificationRuleResponse_lastModifiedTimestamp :: Lens' DescribeNotificationRuleResponse (Maybe UTCTime)

The date and time the notification rule was most recently updated, in timestamp format.

describeNotificationRuleResponse_createdBy :: Lens' DescribeNotificationRuleResponse (Maybe Text)

The name or email alias of the person who created the notification rule.

describeNotificationRuleResponse_detailType :: Lens' DescribeNotificationRuleResponse (Maybe DetailType)

The level of detail included in the notifications for this resource. BASIC will include only the contents of the event as it would appear in AWS CloudWatch. FULL will include any supplemental information provided by AWS CodeStar Notifications and/or the service for the resource for which the notification is created.

describeNotificationRuleResponse_name :: Lens' DescribeNotificationRuleResponse (Maybe Text)

The name of the notification rule.

describeNotificationRuleResponse_targets :: Lens' DescribeNotificationRuleResponse (Maybe [TargetSummary])

A list of the SNS topics associated with the notification rule.

describeNotificationRuleResponse_resource :: Lens' DescribeNotificationRuleResponse (Maybe Text)

The Amazon Resource Name (ARN) of the resource associated with the notification rule.

describeNotificationRuleResponse_createdTimestamp :: Lens' DescribeNotificationRuleResponse (Maybe UTCTime)

The date and time the notification rule was created, in timestamp format.

describeNotificationRuleResponse_tags :: Lens' DescribeNotificationRuleResponse (Maybe (HashMap Text Text))

The tags associated with the notification rule.

describeNotificationRuleResponse_httpStatus :: Lens' DescribeNotificationRuleResponse Int

The response's http status code.

describeNotificationRuleResponse_arn :: Lens' DescribeNotificationRuleResponse Text

The Amazon Resource Name (ARN) of the notification rule.

Types

EventTypeSummary

eventTypeSummary_resourceType :: Lens' EventTypeSummary (Maybe Text)

The resource type of the event.

eventTypeSummary_eventTypeName :: Lens' EventTypeSummary (Maybe Text)

The name of the event.

eventTypeSummary_eventTypeId :: Lens' EventTypeSummary (Maybe Text)

The system-generated ID of the event.

eventTypeSummary_serviceName :: Lens' EventTypeSummary (Maybe Text)

The name of the service for which the event applies.

ListEventTypesFilter

listEventTypesFilter_name :: Lens' ListEventTypesFilter ListEventTypesFilterName

The system-generated name of the filter type you want to filter by.

listEventTypesFilter_value :: Lens' ListEventTypesFilter Text

The name of the resource type (for example, pipeline) or service name (for example, CodePipeline) that you want to filter by.

ListNotificationRulesFilter

listNotificationRulesFilter_name :: Lens' ListNotificationRulesFilter ListNotificationRulesFilterName

The name of the attribute you want to use to filter the returned notification rules.

listNotificationRulesFilter_value :: Lens' ListNotificationRulesFilter Text

The value of the attribute you want to use to filter the returned notification rules. For example, if you specify filtering by RESOURCE in Name, you might specify the ARN of a pipeline in AWS CodePipeline for the value.

ListTargetsFilter

listTargetsFilter_name :: Lens' ListTargetsFilter ListTargetsFilterName

The name of the attribute you want to use to filter the returned targets.

listTargetsFilter_value :: Lens' ListTargetsFilter Text

The value of the attribute you want to use to filter the returned targets. For example, if you specify SNS for the Target type, you could specify an Amazon Resource Name (ARN) for a topic as the value.

NotificationRuleSummary

notificationRuleSummary_arn :: Lens' NotificationRuleSummary (Maybe Text)

The Amazon Resource Name (ARN) of the notification rule.

notificationRuleSummary_id :: Lens' NotificationRuleSummary (Maybe Text)

The unique ID of the notification rule.

Target

target_targetType :: Lens' Target (Maybe Text)

The target type. Can be an Amazon SNS topic.

target_targetAddress :: Lens' Target (Maybe Text)

The Amazon Resource Name (ARN) of the SNS topic.

TargetSummary

targetSummary_targetType :: Lens' TargetSummary (Maybe Text)

The type of the target (for example, SNS).

targetSummary_targetAddress :: Lens' TargetSummary (Maybe Text)

The Amazon Resource Name (ARN) of the SNS topic.

targetSummary_targetStatus :: Lens' TargetSummary (Maybe TargetStatus)

The status of the target.