Amazonka.CognitoIdentity.DeleteIdentities

Description

Deletes identities from an identity pool. You can specify a list of 1-60 identities that you want to delete.

You must use AWS Developer credentials to call this API.

Creating a Request

data DeleteIdentities

Input to the DeleteIdentities action.

See: newDeleteIdentities smart constructor.

Constructors

DeleteIdentities'

Fields

identityIdsToDelete :: NonEmpty Text

A list of 1-60 identities that you want to delete.

newDeleteIdentities

Arguments

:: NonEmpty Text
$sel:identityIdsToDelete:DeleteIdentities'

-> DeleteIdentities

Create a value of DeleteIdentities with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:identityIdsToDelete:DeleteIdentities', deleteIdentities_identityIdsToDelete - A list of 1-60 identities that you want to delete.

Request Lenses

deleteIdentities_identityIdsToDelete :: Lens' DeleteIdentities (NonEmpty Text)

A list of 1-60 identities that you want to delete.

Destructuring the Response

data DeleteIdentitiesResponse

Returned in response to a successful DeleteIdentities operation.

See: newDeleteIdentitiesResponse smart constructor.

Constructors

DeleteIdentitiesResponse'

Fields

unprocessedIdentityIds :: Maybe [UnprocessedIdentityId]

An array of UnprocessedIdentityId objects, each of which contains an ErrorCode and IdentityId.

httpStatus :: Int

The response's http status code.

newDeleteIdentitiesResponse

Arguments

:: Int
$sel:httpStatus:DeleteIdentitiesResponse'

-> DeleteIdentitiesResponse

Create a value of DeleteIdentitiesResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:unprocessedIdentityIds:DeleteIdentitiesResponse', deleteIdentitiesResponse_unprocessedIdentityIds - An array of UnprocessedIdentityId objects, each of which contains an ErrorCode and IdentityId.

$sel:httpStatus:DeleteIdentitiesResponse', deleteIdentitiesResponse_httpStatus - The response's http status code.

Response Lenses

deleteIdentitiesResponse_unprocessedIdentityIds :: Lens' DeleteIdentitiesResponse (Maybe [UnprocessedIdentityId])

An array of UnprocessedIdentityId objects, each of which contains an ErrorCode and IdentityId.

deleteIdentitiesResponse_httpStatus :: Lens' DeleteIdentitiesResponse Int

The response's http status code.