Amazonka.CognitoIdentity.GetOpenIdToken

Description

Gets an OpenID token, using a known Cognito ID. This known Cognito ID is returned by GetId. You can optionally add additional logins for the identity. Supplying multiple logins creates an implicit link.

The OpenID token is valid for 10 minutes.

This is a public API. You do not need any credentials to call this API.

Creating a Request

data GetOpenIdToken

Input to the GetOpenIdToken action.

See: newGetOpenIdToken smart constructor.

Constructors

GetOpenIdToken'

Fields

logins :: Maybe (HashMap Text Text)

A set of optional name-value pairs that map provider names to provider tokens. When using graph.facebook.com and www.amazon.com, supply the access_token returned from the provider's authflow. For accounts.google.com, an Amazon Cognito user pool provider, or any other OpenID Connect provider, always include the id_token.

identityId :: Text

A unique identifier in the format REGION:GUID.

newGetOpenIdToken

Arguments

:: Text
$sel:identityId:GetOpenIdToken'

-> GetOpenIdToken

Create a value of GetOpenIdToken with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:logins:GetOpenIdToken', getOpenIdToken_logins - A set of optional name-value pairs that map provider names to provider tokens. When using graph.facebook.com and www.amazon.com, supply the access_token returned from the provider's authflow. For accounts.google.com, an Amazon Cognito user pool provider, or any other OpenID Connect provider, always include the id_token.

$sel:identityId:GetOpenIdToken', getOpenIdToken_identityId - A unique identifier in the format REGION:GUID.

Request Lenses

getOpenIdToken_logins :: Lens' GetOpenIdToken (Maybe (HashMap Text Text))

A set of optional name-value pairs that map provider names to provider tokens. When using graph.facebook.com and www.amazon.com, supply the access_token returned from the provider's authflow. For accounts.google.com, an Amazon Cognito user pool provider, or any other OpenID Connect provider, always include the id_token.

getOpenIdToken_identityId :: Lens' GetOpenIdToken Text

A unique identifier in the format REGION:GUID.

Destructuring the Response

data GetOpenIdTokenResponse

Returned in response to a successful GetOpenIdToken request.

See: newGetOpenIdTokenResponse smart constructor.

Constructors

GetOpenIdTokenResponse'

Fields

token :: Maybe Text

An OpenID token, valid for 10 minutes.

identityId :: Maybe Text

A unique identifier in the format REGION:GUID. Note that the IdentityId returned may not match the one passed on input.

httpStatus :: Int

The response's http status code.

newGetOpenIdTokenResponse

Arguments

:: Int
$sel:httpStatus:GetOpenIdTokenResponse'

-> GetOpenIdTokenResponse

Create a value of GetOpenIdTokenResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:token:GetOpenIdTokenResponse', getOpenIdTokenResponse_token - An OpenID token, valid for 10 minutes.

$sel:identityId:GetOpenIdToken', getOpenIdTokenResponse_identityId - A unique identifier in the format REGION:GUID. Note that the IdentityId returned may not match the one passed on input.

$sel:httpStatus:GetOpenIdTokenResponse', getOpenIdTokenResponse_httpStatus - The response's http status code.

Response Lenses

getOpenIdTokenResponse_token :: Lens' GetOpenIdTokenResponse (Maybe Text)

An OpenID token, valid for 10 minutes.

getOpenIdTokenResponse_identityId :: Lens' GetOpenIdTokenResponse (Maybe Text)

A unique identifier in the format REGION:GUID. Note that the IdentityId returned may not match the one passed on input.

getOpenIdTokenResponse_httpStatus :: Lens' GetOpenIdTokenResponse Int

The response's http status code.