Amazonka.CognitoIdentity.GetPrincipalTagAttributeMap

Description

Use GetPrincipalTagAttributeMap to list all mappings between PrincipalTags and user attributes.

Creating a Request

data GetPrincipalTagAttributeMap

See: newGetPrincipalTagAttributeMap smart constructor.

Constructors

GetPrincipalTagAttributeMap'

Fields

identityPoolId :: Text

You can use this operation to get the ID of the Identity Pool you setup attribute mappings for.

identityProviderName :: Text

You can use this operation to get the provider name.

newGetPrincipalTagAttributeMap

Arguments

:: Text
$sel:identityPoolId:GetPrincipalTagAttributeMap'

-> Text
$sel:identityProviderName:GetPrincipalTagAttributeMap'

-> GetPrincipalTagAttributeMap

Create a value of GetPrincipalTagAttributeMap with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:identityPoolId:GetPrincipalTagAttributeMap', getPrincipalTagAttributeMap_identityPoolId - You can use this operation to get the ID of the Identity Pool you setup attribute mappings for.

$sel:identityProviderName:GetPrincipalTagAttributeMap', getPrincipalTagAttributeMap_identityProviderName - You can use this operation to get the provider name.

Request Lenses

getPrincipalTagAttributeMap_identityPoolId :: Lens' GetPrincipalTagAttributeMap Text

You can use this operation to get the ID of the Identity Pool you setup attribute mappings for.

getPrincipalTagAttributeMap_identityProviderName :: Lens' GetPrincipalTagAttributeMap Text

You can use this operation to get the provider name.

Destructuring the Response

data GetPrincipalTagAttributeMapResponse

See: newGetPrincipalTagAttributeMapResponse smart constructor.

Constructors

GetPrincipalTagAttributeMapResponse'

Fields

identityPoolId :: Maybe Text

You can use this operation to get the ID of the Identity Pool you setup attribute mappings for.

identityProviderName :: Maybe Text

You can use this operation to get the provider name.

principalTags :: Maybe (HashMap Text Text)

You can use this operation to add principal tags. The PrincipalTagsoperation enables you to reference user attributes in your IAM permissions policy.

useDefaults :: Maybe Bool

You can use this operation to list

httpStatus :: Int

The response's http status code.

newGetPrincipalTagAttributeMapResponse

Arguments

:: Int
$sel:httpStatus:GetPrincipalTagAttributeMapResponse'

-> GetPrincipalTagAttributeMapResponse

Create a value of GetPrincipalTagAttributeMapResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:identityPoolId:GetPrincipalTagAttributeMap', getPrincipalTagAttributeMapResponse_identityPoolId - You can use this operation to get the ID of the Identity Pool you setup attribute mappings for.

$sel:identityProviderName:GetPrincipalTagAttributeMap', getPrincipalTagAttributeMapResponse_identityProviderName - You can use this operation to get the provider name.

$sel:principalTags:GetPrincipalTagAttributeMapResponse', getPrincipalTagAttributeMapResponse_principalTags - You can use this operation to add principal tags. The PrincipalTagsoperation enables you to reference user attributes in your IAM permissions policy.

$sel:useDefaults:GetPrincipalTagAttributeMapResponse', getPrincipalTagAttributeMapResponse_useDefaults - You can use this operation to list

$sel:httpStatus:GetPrincipalTagAttributeMapResponse', getPrincipalTagAttributeMapResponse_httpStatus - The response's http status code.

Response Lenses

getPrincipalTagAttributeMapResponse_identityPoolId :: Lens' GetPrincipalTagAttributeMapResponse (Maybe Text)

You can use this operation to get the ID of the Identity Pool you setup attribute mappings for.

getPrincipalTagAttributeMapResponse_identityProviderName :: Lens' GetPrincipalTagAttributeMapResponse (Maybe Text)

You can use this operation to get the provider name.

getPrincipalTagAttributeMapResponse_principalTags :: Lens' GetPrincipalTagAttributeMapResponse (Maybe (HashMap Text Text))

You can use this operation to add principal tags. The PrincipalTagsoperation enables you to reference user attributes in your IAM permissions policy.

getPrincipalTagAttributeMapResponse_useDefaults :: Lens' GetPrincipalTagAttributeMapResponse (Maybe Bool)

You can use this operation to list

getPrincipalTagAttributeMapResponse_httpStatus :: Lens' GetPrincipalTagAttributeMapResponse Int

The response's http status code.