Amazonka.CognitoIdentity.ListIdentityPools

Description

Lists all of the Cognito identity pools registered for your account.

You must use AWS Developer credentials to call this API.

This operation returns paginated results.

Creating a Request

data ListIdentityPools

Input to the ListIdentityPools action.

See: newListIdentityPools smart constructor.

Constructors

ListIdentityPools'

Fields

nextToken :: Maybe Text

A pagination token.

maxResults :: Natural

The maximum number of identities to return.

newListIdentityPools

Arguments

:: Natural
$sel:maxResults:ListIdentityPools'

-> ListIdentityPools

Create a value of ListIdentityPools with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:nextToken:ListIdentityPools', listIdentityPools_nextToken - A pagination token.

$sel:maxResults:ListIdentityPools', listIdentityPools_maxResults - The maximum number of identities to return.

Request Lenses

listIdentityPools_nextToken :: Lens' ListIdentityPools (Maybe Text)

A pagination token.

listIdentityPools_maxResults :: Lens' ListIdentityPools Natural

The maximum number of identities to return.

Destructuring the Response

data ListIdentityPoolsResponse

The result of a successful ListIdentityPools action.

See: newListIdentityPoolsResponse smart constructor.

Constructors

ListIdentityPoolsResponse'

Fields

identityPools :: Maybe [IdentityPoolShortDescription]

The identity pools returned by the ListIdentityPools action.

nextToken :: Maybe Text

A pagination token.

httpStatus :: Int

The response's http status code.

newListIdentityPoolsResponse

Arguments

:: Int
$sel:httpStatus:ListIdentityPoolsResponse'

-> ListIdentityPoolsResponse

Create a value of ListIdentityPoolsResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:identityPools:ListIdentityPoolsResponse', listIdentityPoolsResponse_identityPools - The identity pools returned by the ListIdentityPools action.

$sel:nextToken:ListIdentityPools', listIdentityPoolsResponse_nextToken - A pagination token.

$sel:httpStatus:ListIdentityPoolsResponse', listIdentityPoolsResponse_httpStatus - The response's http status code.

Response Lenses

listIdentityPoolsResponse_identityPools :: Lens' ListIdentityPoolsResponse (Maybe [IdentityPoolShortDescription])

The identity pools returned by the ListIdentityPools action.

listIdentityPoolsResponse_nextToken :: Lens' ListIdentityPoolsResponse (Maybe Text)

A pagination token.

listIdentityPoolsResponse_httpStatus :: Lens' ListIdentityPoolsResponse Int

The response's http status code.