Amazonka.CognitoIdentity.LookupDeveloperIdentity

Description

Retrieves the IdentityID associated with a DeveloperUserIdentifier or the list of DeveloperUserIdentifier values associated with an IdentityId for an existing identity. Either IdentityID or DeveloperUserIdentifier must not be null. If you supply only one of these values, the other value will be searched in the database and returned as a part of the response. If you supply both, DeveloperUserIdentifier will be matched against IdentityID. If the values are verified against the database, the response returns both values and is the same as the request. Otherwise a ResourceConflictException is thrown.

LookupDeveloperIdentity is intended for low-throughput control plane operations: for example, to enable customer service to locate an identity ID by username. If you are using it for higher-volume operations such as user authentication, your requests are likely to be throttled. GetOpenIdTokenForDeveloperIdentity is a better option for higher-volume operations for user authentication.

You must use AWS Developer credentials to call this API.

Creating a Request

data LookupDeveloperIdentity

Input to the LookupDeveloperIdentityInput action.

See: newLookupDeveloperIdentity smart constructor.

Constructors

LookupDeveloperIdentity'

Fields

developerUserIdentifier :: Maybe Text

A unique ID used by your backend authentication process to identify a user. Typically, a developer identity provider would issue many developer user identifiers, in keeping with the number of users.

nextToken :: Maybe Text

A pagination token. The first call you make will have NextToken set to null. After that the service will return NextToken values as needed. For example, let's say you make a request with MaxResults set to 10, and there are 20 matches in the database. The service will return a pagination token as a part of the response. This token can be used to call the API again and get results starting from the 11th match.

identityId :: Maybe Text

A unique identifier in the format REGION:GUID.

maxResults :: Maybe Natural

The maximum number of identities to return.

identityPoolId :: Text

An identity pool ID in the format REGION:GUID.

newLookupDeveloperIdentity

Arguments

:: Text
$sel:identityPoolId:LookupDeveloperIdentity'

-> LookupDeveloperIdentity

Create a value of LookupDeveloperIdentity with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:developerUserIdentifier:LookupDeveloperIdentity', lookupDeveloperIdentity_developerUserIdentifier - A unique ID used by your backend authentication process to identify a user. Typically, a developer identity provider would issue many developer user identifiers, in keeping with the number of users.

$sel:nextToken:LookupDeveloperIdentity', lookupDeveloperIdentity_nextToken - A pagination token. The first call you make will have NextToken set to null. After that the service will return NextToken values as needed. For example, let's say you make a request with MaxResults set to 10, and there are 20 matches in the database. The service will return a pagination token as a part of the response. This token can be used to call the API again and get results starting from the 11th match.

$sel:identityId:LookupDeveloperIdentity', lookupDeveloperIdentity_identityId - A unique identifier in the format REGION:GUID.

$sel:maxResults:LookupDeveloperIdentity', lookupDeveloperIdentity_maxResults - The maximum number of identities to return.

$sel:identityPoolId:LookupDeveloperIdentity', lookupDeveloperIdentity_identityPoolId - An identity pool ID in the format REGION:GUID.

Request Lenses

lookupDeveloperIdentity_developerUserIdentifier :: Lens' LookupDeveloperIdentity (Maybe Text)

A unique ID used by your backend authentication process to identify a user. Typically, a developer identity provider would issue many developer user identifiers, in keeping with the number of users.

lookupDeveloperIdentity_nextToken :: Lens' LookupDeveloperIdentity (Maybe Text)

A pagination token. The first call you make will have NextToken set to null. After that the service will return NextToken values as needed. For example, let's say you make a request with MaxResults set to 10, and there are 20 matches in the database. The service will return a pagination token as a part of the response. This token can be used to call the API again and get results starting from the 11th match.

lookupDeveloperIdentity_identityId :: Lens' LookupDeveloperIdentity (Maybe Text)

A unique identifier in the format REGION:GUID.

lookupDeveloperIdentity_maxResults :: Lens' LookupDeveloperIdentity (Maybe Natural)

The maximum number of identities to return.

lookupDeveloperIdentity_identityPoolId :: Lens' LookupDeveloperIdentity Text

An identity pool ID in the format REGION:GUID.

Destructuring the Response

data LookupDeveloperIdentityResponse

Returned in response to a successful LookupDeveloperIdentity action.

See: newLookupDeveloperIdentityResponse smart constructor.

Constructors

LookupDeveloperIdentityResponse'

Fields

nextToken :: Maybe Text

A pagination token. The first call you make will have NextToken set to null. After that the service will return NextToken values as needed. For example, let's say you make a request with MaxResults set to 10, and there are 20 matches in the database. The service will return a pagination token as a part of the response. This token can be used to call the API again and get results starting from the 11th match.

identityId :: Maybe Text

A unique identifier in the format REGION:GUID.

developerUserIdentifierList :: Maybe [Text]

This is the list of developer user identifiers associated with an identity ID. Cognito supports the association of multiple developer user identifiers with an identity ID.

httpStatus :: Int

The response's http status code.

newLookupDeveloperIdentityResponse

Arguments

:: Int
$sel:httpStatus:LookupDeveloperIdentityResponse'

-> LookupDeveloperIdentityResponse

Create a value of LookupDeveloperIdentityResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:nextToken:LookupDeveloperIdentity', lookupDeveloperIdentityResponse_nextToken - A pagination token. The first call you make will have NextToken set to null. After that the service will return NextToken values as needed. For example, let's say you make a request with MaxResults set to 10, and there are 20 matches in the database. The service will return a pagination token as a part of the response. This token can be used to call the API again and get results starting from the 11th match.

$sel:identityId:LookupDeveloperIdentity', lookupDeveloperIdentityResponse_identityId - A unique identifier in the format REGION:GUID.

$sel:developerUserIdentifierList:LookupDeveloperIdentityResponse', lookupDeveloperIdentityResponse_developerUserIdentifierList - This is the list of developer user identifiers associated with an identity ID. Cognito supports the association of multiple developer user identifiers with an identity ID.

$sel:httpStatus:LookupDeveloperIdentityResponse', lookupDeveloperIdentityResponse_httpStatus - The response's http status code.

Response Lenses

lookupDeveloperIdentityResponse_nextToken :: Lens' LookupDeveloperIdentityResponse (Maybe Text)

A pagination token. The first call you make will have NextToken set to null. After that the service will return NextToken values as needed. For example, let's say you make a request with MaxResults set to 10, and there are 20 matches in the database. The service will return a pagination token as a part of the response. This token can be used to call the API again and get results starting from the 11th match.

lookupDeveloperIdentityResponse_identityId :: Lens' LookupDeveloperIdentityResponse (Maybe Text)

A unique identifier in the format REGION:GUID.

lookupDeveloperIdentityResponse_developerUserIdentifierList :: Lens' LookupDeveloperIdentityResponse (Maybe [Text])

This is the list of developer user identifiers associated with an identity ID. Cognito supports the association of multiple developer user identifiers with an identity ID.

lookupDeveloperIdentityResponse_httpStatus :: Lens' LookupDeveloperIdentityResponse Int

The response's http status code.