Amazonka.CognitoIdentity.Types.IdentityPool

Description

Documentation

data IdentityPool

An object representing an Amazon Cognito identity pool.

See: newIdentityPool smart constructor.

Constructors

IdentityPool'

Fields

samlProviderARNs :: Maybe [Text]

An array of Amazon Resource Names (ARNs) of the SAML provider for your identity pool.

supportedLoginProviders :: Maybe (HashMap Text Text)

Optional key:value pairs mapping provider names to provider app IDs.

allowClassicFlow :: Maybe Bool

Enables or disables the Basic (Classic) authentication flow. For more information, see Identity Pools (Federated Identities) Authentication Flow in the Amazon Cognito Developer Guide.

developerProviderName :: Maybe Text

The "domain" by which Cognito will refer to your users.

identityPoolTags :: Maybe (HashMap Text Text)

The tags that are assigned to the identity pool. A tag is a label that you can apply to identity pools to categorize and manage them in different ways, such as by purpose, owner, environment, or other criteria.

openIdConnectProviderARNs :: Maybe [Text]

The ARNs of the OpenID Connect providers.

cognitoIdentityProviders :: Maybe [CognitoIdentityProvider]

A list representing an Amazon Cognito user pool and its client ID.

identityPoolId :: Text

An identity pool ID in the format REGION:GUID.

identityPoolName :: Text

A string that you provide.

allowUnauthenticatedIdentities :: Bool

TRUE if the identity pool supports unauthenticated logins.

newIdentityPool

Arguments

:: Text
$sel:identityPoolId:IdentityPool'

-> Text
$sel:identityPoolName:IdentityPool'

-> Bool
$sel:allowUnauthenticatedIdentities:IdentityPool'

-> IdentityPool

Create a value of IdentityPool with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:samlProviderARNs:IdentityPool', identityPool_samlProviderARNs - An array of Amazon Resource Names (ARNs) of the SAML provider for your identity pool.

$sel:supportedLoginProviders:IdentityPool', identityPool_supportedLoginProviders - Optional key:value pairs mapping provider names to provider app IDs.

$sel:allowClassicFlow:IdentityPool', identityPool_allowClassicFlow - Enables or disables the Basic (Classic) authentication flow. For more information, see Identity Pools (Federated Identities) Authentication Flow in the Amazon Cognito Developer Guide.

$sel:developerProviderName:IdentityPool', identityPool_developerProviderName - The "domain" by which Cognito will refer to your users.

$sel:identityPoolTags:IdentityPool', identityPool_identityPoolTags - The tags that are assigned to the identity pool. A tag is a label that you can apply to identity pools to categorize and manage them in different ways, such as by purpose, owner, environment, or other criteria.

$sel:openIdConnectProviderARNs:IdentityPool', identityPool_openIdConnectProviderARNs - The ARNs of the OpenID Connect providers.

$sel:cognitoIdentityProviders:IdentityPool', identityPool_cognitoIdentityProviders - A list representing an Amazon Cognito user pool and its client ID.

$sel:identityPoolId:IdentityPool', identityPool_identityPoolId - An identity pool ID in the format REGION:GUID.

$sel:identityPoolName:IdentityPool', identityPool_identityPoolName - A string that you provide.

$sel:allowUnauthenticatedIdentities:IdentityPool', identityPool_allowUnauthenticatedIdentities - TRUE if the identity pool supports unauthenticated logins.

identityPool_samlProviderARNs :: Lens' IdentityPool (Maybe [Text])

An array of Amazon Resource Names (ARNs) of the SAML provider for your identity pool.

identityPool_supportedLoginProviders :: Lens' IdentityPool (Maybe (HashMap Text Text))

Optional key:value pairs mapping provider names to provider app IDs.

identityPool_allowClassicFlow :: Lens' IdentityPool (Maybe Bool)

Enables or disables the Basic (Classic) authentication flow. For more information, see Identity Pools (Federated Identities) Authentication Flow in the Amazon Cognito Developer Guide.

identityPool_developerProviderName :: Lens' IdentityPool (Maybe Text)

The "domain" by which Cognito will refer to your users.

identityPool_identityPoolTags :: Lens' IdentityPool (Maybe (HashMap Text Text))

The tags that are assigned to the identity pool. A tag is a label that you can apply to identity pools to categorize and manage them in different ways, such as by purpose, owner, environment, or other criteria.

identityPool_openIdConnectProviderARNs :: Lens' IdentityPool (Maybe [Text])

The ARNs of the OpenID Connect providers.

identityPool_cognitoIdentityProviders :: Lens' IdentityPool (Maybe [CognitoIdentityProvider])

A list representing an Amazon Cognito user pool and its client ID.

identityPool_identityPoolId :: Lens' IdentityPool Text

An identity pool ID in the format REGION:GUID.

identityPool_identityPoolName :: Lens' IdentityPool Text

A string that you provide.

identityPool_allowUnauthenticatedIdentities :: Lens' IdentityPool Bool

TRUE if the identity pool supports unauthenticated logins.