Amazonka.CognitoIdentity.Types.MappingRule

Description

Documentation

data MappingRule

A rule that maps a claim name, a claim value, and a match type to a role ARN.

See: newMappingRule smart constructor.

Constructors

MappingRule'

Fields

claim :: Text

The claim name that must be present in the token, for example, "isAdmin" or "paid".

matchType :: MappingRuleMatchType

The match condition that specifies how closely the claim value in the IdP token must match Value.

value :: Text

A brief string that the claim must match, for example, "paid" or "yes".

roleARN :: Text

The role ARN.

newMappingRule

Arguments

:: Text
$sel:claim:MappingRule'

-> MappingRuleMatchType
$sel:matchType:MappingRule'

-> Text
$sel:value:MappingRule'

-> Text
$sel:roleARN:MappingRule'

-> MappingRule

Create a value of MappingRule with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:claim:MappingRule', mappingRule_claim - The claim name that must be present in the token, for example, "isAdmin" or "paid".

$sel:matchType:MappingRule', mappingRule_matchType - The match condition that specifies how closely the claim value in the IdP token must match Value.

$sel:value:MappingRule', mappingRule_value - A brief string that the claim must match, for example, "paid" or "yes".

$sel:roleARN:MappingRule', mappingRule_roleARN - The role ARN.

mappingRule_claim :: Lens' MappingRule Text

The claim name that must be present in the token, for example, "isAdmin" or "paid".

mappingRule_matchType :: Lens' MappingRule MappingRuleMatchType

The match condition that specifies how closely the claim value in the IdP token must match Value.

mappingRule_value :: Lens' MappingRule Text

A brief string that the claim must match, for example, "paid" or "yes".

mappingRule_roleARN :: Lens' MappingRule Text

The role ARN.