Amazonka.CognitoIdentityProvider.AdminListUserAuthEvents

Description

Lists a history of user activity and any risks detected as part of Amazon Cognito advanced security.

This operation returns paginated results.

Creating a Request

data AdminListUserAuthEvents

See: newAdminListUserAuthEvents smart constructor.

Constructors

AdminListUserAuthEvents'

Fields

nextToken :: Maybe Text

A pagination token.

maxResults :: Maybe Natural

The maximum number of authentication events to return.

userPoolId :: Text

The user pool ID.

username :: Sensitive Text

The user pool username or an alias.

newAdminListUserAuthEvents

Arguments

:: Text
$sel:userPoolId:AdminListUserAuthEvents'

-> Text
$sel:username:AdminListUserAuthEvents'

-> AdminListUserAuthEvents

Create a value of AdminListUserAuthEvents with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:nextToken:AdminListUserAuthEvents', adminListUserAuthEvents_nextToken - A pagination token.

$sel:maxResults:AdminListUserAuthEvents', adminListUserAuthEvents_maxResults - The maximum number of authentication events to return.

$sel:userPoolId:AdminListUserAuthEvents', adminListUserAuthEvents_userPoolId - The user pool ID.

$sel:username:AdminListUserAuthEvents', adminListUserAuthEvents_username - The user pool username or an alias.

Request Lenses

adminListUserAuthEvents_nextToken :: Lens' AdminListUserAuthEvents (Maybe Text)

A pagination token.

adminListUserAuthEvents_maxResults :: Lens' AdminListUserAuthEvents (Maybe Natural)

The maximum number of authentication events to return.

adminListUserAuthEvents_userPoolId :: Lens' AdminListUserAuthEvents Text

The user pool ID.

adminListUserAuthEvents_username :: Lens' AdminListUserAuthEvents Text

The user pool username or an alias.

Destructuring the Response

data AdminListUserAuthEventsResponse

See: newAdminListUserAuthEventsResponse smart constructor.

Constructors

AdminListUserAuthEventsResponse'

Fields

nextToken :: Maybe Text

A pagination token.

authEvents :: Maybe [AuthEventType]

The response object. It includes the EventID, EventType, CreationDate, EventRisk, and EventResponse.

httpStatus :: Int

The response's http status code.

newAdminListUserAuthEventsResponse

Arguments

:: Int
$sel:httpStatus:AdminListUserAuthEventsResponse'

-> AdminListUserAuthEventsResponse

Create a value of AdminListUserAuthEventsResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:nextToken:AdminListUserAuthEvents', adminListUserAuthEventsResponse_nextToken - A pagination token.

$sel:authEvents:AdminListUserAuthEventsResponse', adminListUserAuthEventsResponse_authEvents - The response object. It includes the EventID, EventType, CreationDate, EventRisk, and EventResponse.

$sel:httpStatus:AdminListUserAuthEventsResponse', adminListUserAuthEventsResponse_httpStatus - The response's http status code.

Response Lenses

adminListUserAuthEventsResponse_nextToken :: Lens' AdminListUserAuthEventsResponse (Maybe Text)

A pagination token.

adminListUserAuthEventsResponse_authEvents :: Lens' AdminListUserAuthEventsResponse (Maybe [AuthEventType])

The response object. It includes the EventID, EventType, CreationDate, EventRisk, and EventResponse.

adminListUserAuthEventsResponse_httpStatus :: Lens' AdminListUserAuthEventsResponse Int

The response's http status code.