Amazonka.CognitoIdentityProvider.ConfirmForgotPassword

Description

Allows a user to enter a confirmation code to reset a forgotten password.

Creating a Request

data ConfirmForgotPassword

The request representing the confirmation for a password reset.

See: newConfirmForgotPassword smart constructor.

Constructors

ConfirmForgotPassword'

Fields

clientMetadata :: Maybe (HashMap Text Text)

A map of custom key-value pairs that you can provide as input for any custom workflows that this action triggers.

You create custom workflows by assigning Lambda functions to user pool triggers. When you use the ConfirmForgotPassword API action, Amazon Cognito invokes the function that is assigned to the post confirmation trigger. When Amazon Cognito invokes this function, it passes a JSON payload, which the function receives as input. This payload contains a clientMetadata attribute, which provides the data that you assigned to the ClientMetadata parameter in your ConfirmForgotPassword request. In your function code in Lambda, you can process the clientMetadata value to enhance your workflow for your specific needs.

For more information, see Customizing User Pool Workflows with Lambda Triggers in the Amazon Cognito Developer Guide.

Take the following limitations into consideration when you use the ClientMetadata parameter:

Amazon Cognito does not store the ClientMetadata value. This data is available only to Lambda triggers that are assigned to a user pool to support custom workflows. If your user pool configuration does not include triggers, the ClientMetadata parameter serves no purpose.

Amazon Cognito does not validate the ClientMetadata value.

Amazon Cognito does not encrypt the the ClientMetadata value, so don't use it to provide sensitive information.

analyticsMetadata :: Maybe AnalyticsMetadataType

The Amazon Pinpoint analytics metadata for collecting metrics for ConfirmForgotPassword calls.

userContextData :: Maybe UserContextDataType

Contextual data such as the user's device fingerprint, IP address, or location used for evaluating the risk of an unexpected event by Amazon Cognito advanced security.

secretHash :: Maybe (Sensitive Text)

A keyed-hash message authentication code (HMAC) calculated using the secret key of a user pool client and username plus the client ID in the message.

clientId :: Sensitive Text

The app client ID of the app associated with the user pool.

username :: Sensitive Text

The user name of the user for whom you want to enter a code to retrieve a forgotten password.

confirmationCode :: Text

The confirmation code sent by a user's request to retrieve a forgotten password. For more information, see ForgotPassword.

password :: Sensitive Text

The password sent by a user's request to retrieve a forgotten password.

newConfirmForgotPassword

Arguments

:: Text
$sel:clientId:ConfirmForgotPassword'

-> Text
$sel:username:ConfirmForgotPassword'

-> Text
$sel:confirmationCode:ConfirmForgotPassword'

-> Text
$sel:password:ConfirmForgotPassword'

-> ConfirmForgotPassword

Create a value of ConfirmForgotPassword with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:clientMetadata:ConfirmForgotPassword', confirmForgotPassword_clientMetadata - A map of custom key-value pairs that you can provide as input for any custom workflows that this action triggers.

You create custom workflows by assigning Lambda functions to user pool triggers. When you use the ConfirmForgotPassword API action, Amazon Cognito invokes the function that is assigned to the post confirmation trigger. When Amazon Cognito invokes this function, it passes a JSON payload, which the function receives as input. This payload contains a clientMetadata attribute, which provides the data that you assigned to the ClientMetadata parameter in your ConfirmForgotPassword request. In your function code in Lambda, you can process the clientMetadata value to enhance your workflow for your specific needs.

For more information, see Customizing User Pool Workflows with Lambda Triggers in the Amazon Cognito Developer Guide.

Take the following limitations into consideration when you use the ClientMetadata parameter:

Amazon Cognito does not store the ClientMetadata value. This data is available only to Lambda triggers that are assigned to a user pool to support custom workflows. If your user pool configuration does not include triggers, the ClientMetadata parameter serves no purpose.

Amazon Cognito does not validate the ClientMetadata value.

Amazon Cognito does not encrypt the the ClientMetadata value, so don't use it to provide sensitive information.

$sel:analyticsMetadata:ConfirmForgotPassword', confirmForgotPassword_analyticsMetadata - The Amazon Pinpoint analytics metadata for collecting metrics for ConfirmForgotPassword calls.

$sel:userContextData:ConfirmForgotPassword', confirmForgotPassword_userContextData - Contextual data such as the user's device fingerprint, IP address, or location used for evaluating the risk of an unexpected event by Amazon Cognito advanced security.

$sel:secretHash:ConfirmForgotPassword', confirmForgotPassword_secretHash - A keyed-hash message authentication code (HMAC) calculated using the secret key of a user pool client and username plus the client ID in the message.

$sel:clientId:ConfirmForgotPassword', confirmForgotPassword_clientId - The app client ID of the app associated with the user pool.

$sel:username:ConfirmForgotPassword', confirmForgotPassword_username - The user name of the user for whom you want to enter a code to retrieve a forgotten password.

$sel:confirmationCode:ConfirmForgotPassword', confirmForgotPassword_confirmationCode - The confirmation code sent by a user's request to retrieve a forgotten password. For more information, see ForgotPassword.

$sel:password:ConfirmForgotPassword', confirmForgotPassword_password - The password sent by a user's request to retrieve a forgotten password.

Request Lenses

confirmForgotPassword_clientMetadata :: Lens' ConfirmForgotPassword (Maybe (HashMap Text Text))

A map of custom key-value pairs that you can provide as input for any custom workflows that this action triggers.

You create custom workflows by assigning Lambda functions to user pool triggers. When you use the ConfirmForgotPassword API action, Amazon Cognito invokes the function that is assigned to the post confirmation trigger. When Amazon Cognito invokes this function, it passes a JSON payload, which the function receives as input. This payload contains a clientMetadata attribute, which provides the data that you assigned to the ClientMetadata parameter in your ConfirmForgotPassword request. In your function code in Lambda, you can process the clientMetadata value to enhance your workflow for your specific needs.

For more information, see Customizing User Pool Workflows with Lambda Triggers in the Amazon Cognito Developer Guide.

Take the following limitations into consideration when you use the ClientMetadata parameter:

Amazon Cognito does not store the ClientMetadata value. This data is available only to Lambda triggers that are assigned to a user pool to support custom workflows. If your user pool configuration does not include triggers, the ClientMetadata parameter serves no purpose.

Amazon Cognito does not validate the ClientMetadata value.

Amazon Cognito does not encrypt the the ClientMetadata value, so don't use it to provide sensitive information.

confirmForgotPassword_analyticsMetadata :: Lens' ConfirmForgotPassword (Maybe AnalyticsMetadataType)

The Amazon Pinpoint analytics metadata for collecting metrics for ConfirmForgotPassword calls.

confirmForgotPassword_userContextData :: Lens' ConfirmForgotPassword (Maybe UserContextDataType)

Contextual data such as the user's device fingerprint, IP address, or location used for evaluating the risk of an unexpected event by Amazon Cognito advanced security.

confirmForgotPassword_secretHash :: Lens' ConfirmForgotPassword (Maybe Text)

A keyed-hash message authentication code (HMAC) calculated using the secret key of a user pool client and username plus the client ID in the message.

confirmForgotPassword_clientId :: Lens' ConfirmForgotPassword Text

The app client ID of the app associated with the user pool.

confirmForgotPassword_username :: Lens' ConfirmForgotPassword Text

The user name of the user for whom you want to enter a code to retrieve a forgotten password.

confirmForgotPassword_confirmationCode :: Lens' ConfirmForgotPassword Text

The confirmation code sent by a user's request to retrieve a forgotten password. For more information, see ForgotPassword.

confirmForgotPassword_password :: Lens' ConfirmForgotPassword Text

The password sent by a user's request to retrieve a forgotten password.

Destructuring the Response

data ConfirmForgotPasswordResponse

The response from the server that results from a user's request to retrieve a forgotten password.

See: newConfirmForgotPasswordResponse smart constructor.

Constructors

ConfirmForgotPasswordResponse'

Fields

httpStatus :: Int

The response's http status code.

newConfirmForgotPasswordResponse

Arguments

:: Int
$sel:httpStatus:ConfirmForgotPasswordResponse'

-> ConfirmForgotPasswordResponse

Create a value of ConfirmForgotPasswordResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:httpStatus:ConfirmForgotPasswordResponse', confirmForgotPasswordResponse_httpStatus - The response's http status code.

Response Lenses

confirmForgotPasswordResponse_httpStatus :: Lens' ConfirmForgotPasswordResponse Int

The response's http status code.