Amazonka.CognitoIdentityProvider.CreateIdentityProvider

Description

Creates an identity provider for a user pool.

Creating a Request

data CreateIdentityProvider

See: newCreateIdentityProvider smart constructor.

Constructors

CreateIdentityProvider'

Fields

idpIdentifiers :: Maybe [Text]

A list of identity provider identifiers.

attributeMapping :: Maybe (HashMap Text Text)

A mapping of identity provider attributes to standard and custom user pool attributes.

userPoolId :: Text

The user pool ID.

providerName :: Text

The identity provider name.

providerType :: IdentityProviderTypeType

The identity provider type.

providerDetails :: HashMap Text Text

The identity provider details. The following list describes the provider detail keys for each identity provider type.

For Google and Login with Amazon:

client_id

client_secret

authorize_scopes

For Facebook:

client_id

client_secret

authorize_scopes

api_version

For Sign in with Apple:

client_id

team_id

key_id

private_key

authorize_scopes

For OIDC providers:

client_id

client_secret

attributes_request_method

oidc_issuer

authorize_scopes

authorize_url /if not available from discovery URL specified by oidc_issuer key/

token_url /if not available from discovery URL specified by oidc_issuer key/

attributes_url /if not available from discovery URL specified by oidc_issuer key/

jwks_uri /if not available from discovery URL specified by oidc_issuer key/

For SAML providers:

MetadataFile OR MetadataURL

IDPSignout optional

newCreateIdentityProvider

Arguments

:: Text
$sel:userPoolId:CreateIdentityProvider'

-> Text
$sel:providerName:CreateIdentityProvider'

-> IdentityProviderTypeType
$sel:providerType:CreateIdentityProvider'

-> CreateIdentityProvider

Create a value of CreateIdentityProvider with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:idpIdentifiers:CreateIdentityProvider', createIdentityProvider_idpIdentifiers - A list of identity provider identifiers.

$sel:attributeMapping:CreateIdentityProvider', createIdentityProvider_attributeMapping - A mapping of identity provider attributes to standard and custom user pool attributes.

$sel:userPoolId:CreateIdentityProvider', createIdentityProvider_userPoolId - The user pool ID.

$sel:providerName:CreateIdentityProvider', createIdentityProvider_providerName - The identity provider name.

$sel:providerType:CreateIdentityProvider', createIdentityProvider_providerType - The identity provider type.

$sel:providerDetails:CreateIdentityProvider', createIdentityProvider_providerDetails - The identity provider details. The following list describes the provider detail keys for each identity provider type.

For Google and Login with Amazon:

client_id

client_secret

authorize_scopes

For Facebook:

client_id

client_secret

authorize_scopes

api_version

For Sign in with Apple:

client_id

team_id

key_id

private_key

authorize_scopes

For OIDC providers:

client_id

client_secret

attributes_request_method

oidc_issuer

authorize_scopes

authorize_url /if not available from discovery URL specified by oidc_issuer key/

token_url /if not available from discovery URL specified by oidc_issuer key/

attributes_url /if not available from discovery URL specified by oidc_issuer key/

jwks_uri /if not available from discovery URL specified by oidc_issuer key/

For SAML providers:

MetadataFile OR MetadataURL

IDPSignout optional

Request Lenses

createIdentityProvider_idpIdentifiers :: Lens' CreateIdentityProvider (Maybe [Text])

A list of identity provider identifiers.

createIdentityProvider_attributeMapping :: Lens' CreateIdentityProvider (Maybe (HashMap Text Text))

A mapping of identity provider attributes to standard and custom user pool attributes.

createIdentityProvider_userPoolId :: Lens' CreateIdentityProvider Text

The user pool ID.

createIdentityProvider_providerName :: Lens' CreateIdentityProvider Text

The identity provider name.

createIdentityProvider_providerType :: Lens' CreateIdentityProvider IdentityProviderTypeType

The identity provider type.

createIdentityProvider_providerDetails :: Lens' CreateIdentityProvider (HashMap Text Text)

The identity provider details. The following list describes the provider detail keys for each identity provider type.

For Google and Login with Amazon:

client_id

client_secret

authorize_scopes

For Facebook:

client_id

client_secret

authorize_scopes

api_version

For Sign in with Apple:

client_id

team_id

key_id

private_key

authorize_scopes

For OIDC providers:

client_id

client_secret

attributes_request_method

oidc_issuer

authorize_scopes

authorize_url /if not available from discovery URL specified by oidc_issuer key/

token_url /if not available from discovery URL specified by oidc_issuer key/

attributes_url /if not available from discovery URL specified by oidc_issuer key/

jwks_uri /if not available from discovery URL specified by oidc_issuer key/

For SAML providers:

MetadataFile OR MetadataURL

IDPSignout optional

Destructuring the Response

data CreateIdentityProviderResponse

See: newCreateIdentityProviderResponse smart constructor.

Constructors

CreateIdentityProviderResponse'

Fields

httpStatus :: Int

The response's http status code.

identityProvider :: IdentityProviderType

The newly created identity provider object.

newCreateIdentityProviderResponse

Arguments

:: Int
$sel:httpStatus:CreateIdentityProviderResponse'

-> IdentityProviderType
$sel:identityProvider:CreateIdentityProviderResponse'

-> CreateIdentityProviderResponse

Create a value of CreateIdentityProviderResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:httpStatus:CreateIdentityProviderResponse', createIdentityProviderResponse_httpStatus - The response's http status code.

$sel:identityProvider:CreateIdentityProviderResponse', createIdentityProviderResponse_identityProvider - The newly created identity provider object.

Response Lenses

createIdentityProviderResponse_httpStatus :: Lens' CreateIdentityProviderResponse Int

The response's http status code.

createIdentityProviderResponse_identityProvider :: Lens' CreateIdentityProviderResponse IdentityProviderType

The newly created identity provider object.