Amazonka.CognitoIdentityProvider.RespondToAuthChallenge

Description

Responds to the authentication challenge.

This action might generate an SMS text message. Starting June 1, 2021, U.S. telecom carriers require that you register an origination phone number before you can send SMS messages to U.S. phone numbers. If you use SMS text messages in Amazon Cognito, you must register a phone number with Amazon Pinpoint. Cognito will use the the registered number automatically. Otherwise, Cognito users that must receive SMS messages might be unable to sign up, activate their accounts, or sign in.

If you have never used SMS text messages with Amazon Cognito or any other Amazon Web Service, Amazon SNS might place your account in SMS sandbox. In <https:docs.aws.amazon.comsnslatestdgsns-sms-sandbox.html sandbox mode> , you’ll have limitations, such as sending messages to only verified phone numbers. After testing in the sandbox environment, you can move out of the SMS sandbox and into production. For more information, see SMS message settings for Cognito User Pools in the Amazon Cognito Developer Guide.

Creating a Request

data RespondToAuthChallenge

The request to respond to an authentication challenge.

See: newRespondToAuthChallenge smart constructor.

Constructors

RespondToAuthChallenge'

Fields

clientMetadata :: Maybe (HashMap Text Text)

A map of custom key-value pairs that you can provide as input for any custom workflows that this action triggers.

You create custom workflows by assigning Lambda functions to user pool triggers. When you use the RespondToAuthChallenge API action, Amazon Cognito invokes any functions that are assigned to the following triggers: post authentication, pre token generation, /define auth challenge, create auth challenge, and verify auth challenge/. When Amazon Cognito invokes any of these functions, it passes a JSON payload, which the function receives as input. This payload contains a clientMetadata attribute, which provides the data that you assigned to the ClientMetadata parameter in your RespondToAuthChallenge request. In your function code in Lambda, you can process the clientMetadata value to enhance your workflow for your specific needs.

For more information, see Customizing User Pool Workflows with Lambda Triggers in the Amazon Cognito Developer Guide.

Take the following limitations into consideration when you use the ClientMetadata parameter:

Amazon Cognito does not store the ClientMetadata value. This data is available only to Lambda triggers that are assigned to a user pool to support custom workflows. If your user pool configuration does not include triggers, the ClientMetadata parameter serves no purpose.

Amazon Cognito does not validate the ClientMetadata value.

Amazon Cognito does not encrypt the the ClientMetadata value, so don't use it to provide sensitive information.

analyticsMetadata :: Maybe AnalyticsMetadataType

The Amazon Pinpoint analytics metadata for collecting metrics for RespondToAuthChallenge calls.

challengeResponses :: Maybe (HashMap Text Text)

The challenge responses. These are inputs corresponding to the value of ChallengeName, for example:

SECRET_HASH (if app client is configured with client secret) applies to all inputs below (including SOFTWARE_TOKEN_MFA).

SMS_MFA: SMS_MFA_CODE, USERNAME.

PASSWORD_VERIFIER: PASSWORD_CLAIM_SIGNATURE, PASSWORD_CLAIM_SECRET_BLOCK, TIMESTAMP, USERNAME.

NEW_PASSWORD_REQUIRED: NEW_PASSWORD, any other required attributes, USERNAME.

SOFTWARE_TOKEN_MFA: USERNAME and SOFTWARE_TOKEN_MFA_CODE are required attributes.

DEVICE_SRP_AUTH requires USERNAME, DEVICE_KEY, SRP_A (and SECRET_HASH).

DEVICE_PASSWORD_VERIFIER requires everything that PASSWORD_VERIFIER requires plus DEVICE_KEY.

MFA_SETUP requires USERNAME, plus you need to use the session value returned by VerifySoftwareToken in the Session parameter.

userContextData :: Maybe UserContextDataType

Contextual data such as the user's device fingerprint, IP address, or location used for evaluating the risk of an unexpected event by Amazon Cognito advanced security.

session :: Maybe Text

The session which should be passed both ways in challenge-response calls to the service. If InitiateAuth or RespondToAuthChallenge API call determines that the caller needs to go through another challenge, they return a session with other challenge parameters. This session should be passed as it is to the next RespondToAuthChallenge API call.

clientId :: Sensitive Text

The app client ID.

challengeName :: ChallengeNameType

The challenge name. For more information, see InitiateAuth.

ADMIN_NO_SRP_AUTH is not a valid value.

newRespondToAuthChallenge

Arguments

:: Text
$sel:clientId:RespondToAuthChallenge'

-> ChallengeNameType
$sel:challengeName:RespondToAuthChallenge'

-> RespondToAuthChallenge

Create a value of RespondToAuthChallenge with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:clientMetadata:RespondToAuthChallenge', respondToAuthChallenge_clientMetadata - A map of custom key-value pairs that you can provide as input for any custom workflows that this action triggers.

You create custom workflows by assigning Lambda functions to user pool triggers. When you use the RespondToAuthChallenge API action, Amazon Cognito invokes any functions that are assigned to the following triggers: post authentication, pre token generation, /define auth challenge, create auth challenge, and verify auth challenge/. When Amazon Cognito invokes any of these functions, it passes a JSON payload, which the function receives as input. This payload contains a clientMetadata attribute, which provides the data that you assigned to the ClientMetadata parameter in your RespondToAuthChallenge request. In your function code in Lambda, you can process the clientMetadata value to enhance your workflow for your specific needs.

For more information, see Customizing User Pool Workflows with Lambda Triggers in the Amazon Cognito Developer Guide.

Take the following limitations into consideration when you use the ClientMetadata parameter:

Amazon Cognito does not store the ClientMetadata value. This data is available only to Lambda triggers that are assigned to a user pool to support custom workflows. If your user pool configuration does not include triggers, the ClientMetadata parameter serves no purpose.

Amazon Cognito does not validate the ClientMetadata value.

Amazon Cognito does not encrypt the the ClientMetadata value, so don't use it to provide sensitive information.

$sel:analyticsMetadata:RespondToAuthChallenge', respondToAuthChallenge_analyticsMetadata - The Amazon Pinpoint analytics metadata for collecting metrics for RespondToAuthChallenge calls.

$sel:challengeResponses:RespondToAuthChallenge', respondToAuthChallenge_challengeResponses - The challenge responses. These are inputs corresponding to the value of ChallengeName, for example:

SECRET_HASH (if app client is configured with client secret) applies to all inputs below (including SOFTWARE_TOKEN_MFA).

SMS_MFA: SMS_MFA_CODE, USERNAME.

PASSWORD_VERIFIER: PASSWORD_CLAIM_SIGNATURE, PASSWORD_CLAIM_SECRET_BLOCK, TIMESTAMP, USERNAME.

NEW_PASSWORD_REQUIRED: NEW_PASSWORD, any other required attributes, USERNAME.

SOFTWARE_TOKEN_MFA: USERNAME and SOFTWARE_TOKEN_MFA_CODE are required attributes.

DEVICE_SRP_AUTH requires USERNAME, DEVICE_KEY, SRP_A (and SECRET_HASH).

DEVICE_PASSWORD_VERIFIER requires everything that PASSWORD_VERIFIER requires plus DEVICE_KEY.

MFA_SETUP requires USERNAME, plus you need to use the session value returned by VerifySoftwareToken in the Session parameter.

$sel:userContextData:RespondToAuthChallenge', respondToAuthChallenge_userContextData - Contextual data such as the user's device fingerprint, IP address, or location used for evaluating the risk of an unexpected event by Amazon Cognito advanced security.

$sel:session:RespondToAuthChallenge', respondToAuthChallenge_session - The session which should be passed both ways in challenge-response calls to the service. If InitiateAuth or RespondToAuthChallenge API call determines that the caller needs to go through another challenge, they return a session with other challenge parameters. This session should be passed as it is to the next RespondToAuthChallenge API call.

$sel:clientId:RespondToAuthChallenge', respondToAuthChallenge_clientId - The app client ID.

$sel:challengeName:RespondToAuthChallenge', respondToAuthChallenge_challengeName - The challenge name. For more information, see InitiateAuth.

ADMIN_NO_SRP_AUTH is not a valid value.

Request Lenses

respondToAuthChallenge_clientMetadata :: Lens' RespondToAuthChallenge (Maybe (HashMap Text Text))

A map of custom key-value pairs that you can provide as input for any custom workflows that this action triggers.

You create custom workflows by assigning Lambda functions to user pool triggers. When you use the RespondToAuthChallenge API action, Amazon Cognito invokes any functions that are assigned to the following triggers: post authentication, pre token generation, /define auth challenge, create auth challenge, and verify auth challenge/. When Amazon Cognito invokes any of these functions, it passes a JSON payload, which the function receives as input. This payload contains a clientMetadata attribute, which provides the data that you assigned to the ClientMetadata parameter in your RespondToAuthChallenge request. In your function code in Lambda, you can process the clientMetadata value to enhance your workflow for your specific needs.

For more information, see Customizing User Pool Workflows with Lambda Triggers in the Amazon Cognito Developer Guide.

Take the following limitations into consideration when you use the ClientMetadata parameter:

Amazon Cognito does not store the ClientMetadata value. This data is available only to Lambda triggers that are assigned to a user pool to support custom workflows. If your user pool configuration does not include triggers, the ClientMetadata parameter serves no purpose.

Amazon Cognito does not validate the ClientMetadata value.

Amazon Cognito does not encrypt the the ClientMetadata value, so don't use it to provide sensitive information.

respondToAuthChallenge_analyticsMetadata :: Lens' RespondToAuthChallenge (Maybe AnalyticsMetadataType)

The Amazon Pinpoint analytics metadata for collecting metrics for RespondToAuthChallenge calls.

respondToAuthChallenge_challengeResponses :: Lens' RespondToAuthChallenge (Maybe (HashMap Text Text))

The challenge responses. These are inputs corresponding to the value of ChallengeName, for example:

SECRET_HASH (if app client is configured with client secret) applies to all inputs below (including SOFTWARE_TOKEN_MFA).

SMS_MFA: SMS_MFA_CODE, USERNAME.

PASSWORD_VERIFIER: PASSWORD_CLAIM_SIGNATURE, PASSWORD_CLAIM_SECRET_BLOCK, TIMESTAMP, USERNAME.

NEW_PASSWORD_REQUIRED: NEW_PASSWORD, any other required attributes, USERNAME.

SOFTWARE_TOKEN_MFA: USERNAME and SOFTWARE_TOKEN_MFA_CODE are required attributes.

DEVICE_SRP_AUTH requires USERNAME, DEVICE_KEY, SRP_A (and SECRET_HASH).

DEVICE_PASSWORD_VERIFIER requires everything that PASSWORD_VERIFIER requires plus DEVICE_KEY.

MFA_SETUP requires USERNAME, plus you need to use the session value returned by VerifySoftwareToken in the Session parameter.

respondToAuthChallenge_userContextData :: Lens' RespondToAuthChallenge (Maybe UserContextDataType)

Contextual data such as the user's device fingerprint, IP address, or location used for evaluating the risk of an unexpected event by Amazon Cognito advanced security.

respondToAuthChallenge_session :: Lens' RespondToAuthChallenge (Maybe Text)

The session which should be passed both ways in challenge-response calls to the service. If InitiateAuth or RespondToAuthChallenge API call determines that the caller needs to go through another challenge, they return a session with other challenge parameters. This session should be passed as it is to the next RespondToAuthChallenge API call.

respondToAuthChallenge_clientId :: Lens' RespondToAuthChallenge Text

The app client ID.

respondToAuthChallenge_challengeName :: Lens' RespondToAuthChallenge ChallengeNameType

The challenge name. For more information, see InitiateAuth.

ADMIN_NO_SRP_AUTH is not a valid value.

Destructuring the Response

data RespondToAuthChallengeResponse

The response to respond to the authentication challenge.

See: newRespondToAuthChallengeResponse smart constructor.

Constructors

RespondToAuthChallengeResponse'

Fields

challengeName :: Maybe ChallengeNameType

The challenge name. For more information, see InitiateAuth.

challengeParameters :: Maybe (HashMap Text Text)

The challenge parameters. For more information, see InitiateAuth.

authenticationResult :: Maybe AuthenticationResultType

The result returned by the server in response to the request to respond to the authentication challenge.

session :: Maybe Text

The session which should be passed both ways in challenge-response calls to the service. If the caller needs to go through another challenge, they return a session with other challenge parameters. This session should be passed as it is to the next RespondToAuthChallenge API call.

httpStatus :: Int

The response's http status code.

newRespondToAuthChallengeResponse

Arguments

:: Int
$sel:httpStatus:RespondToAuthChallengeResponse'

-> RespondToAuthChallengeResponse

Create a value of RespondToAuthChallengeResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:challengeName:RespondToAuthChallenge', respondToAuthChallengeResponse_challengeName - The challenge name. For more information, see InitiateAuth.

$sel:challengeParameters:RespondToAuthChallengeResponse', respondToAuthChallengeResponse_challengeParameters - The challenge parameters. For more information, see InitiateAuth.

$sel:authenticationResult:RespondToAuthChallengeResponse', respondToAuthChallengeResponse_authenticationResult - The result returned by the server in response to the request to respond to the authentication challenge.

$sel:session:RespondToAuthChallenge', respondToAuthChallengeResponse_session - The session which should be passed both ways in challenge-response calls to the service. If the caller needs to go through another challenge, they return a session with other challenge parameters. This session should be passed as it is to the next RespondToAuthChallenge API call.

$sel:httpStatus:RespondToAuthChallengeResponse', respondToAuthChallengeResponse_httpStatus - The response's http status code.

Response Lenses

respondToAuthChallengeResponse_challengeName :: Lens' RespondToAuthChallengeResponse (Maybe ChallengeNameType)

The challenge name. For more information, see InitiateAuth.

respondToAuthChallengeResponse_challengeParameters :: Lens' RespondToAuthChallengeResponse (Maybe (HashMap Text Text))

The challenge parameters. For more information, see InitiateAuth.

respondToAuthChallengeResponse_authenticationResult :: Lens' RespondToAuthChallengeResponse (Maybe AuthenticationResultType)

The result returned by the server in response to the request to respond to the authentication challenge.

respondToAuthChallengeResponse_session :: Lens' RespondToAuthChallengeResponse (Maybe Text)

The session which should be passed both ways in challenge-response calls to the service. If the caller needs to go through another challenge, they return a session with other challenge parameters. This session should be passed as it is to the next RespondToAuthChallenge API call.

respondToAuthChallengeResponse_httpStatus :: Lens' RespondToAuthChallengeResponse Int

The response's http status code.