Amazonka.CognitoIdentityProvider.Types.CompromisedCredentialsRiskConfigurationType

Description

Documentation

data CompromisedCredentialsRiskConfigurationType

The compromised credentials risk configuration type.

See: newCompromisedCredentialsRiskConfigurationType smart constructor.

Constructors

CompromisedCredentialsRiskConfigurationType'

Fields

eventFilter :: Maybe [EventFilterType]

Perform the action for these events. The default is to perform all events if no event filter is specified.

actions :: CompromisedCredentialsActionsType

The compromised credentials risk configuration actions.

newCompromisedCredentialsRiskConfigurationType

Arguments

:: CompromisedCredentialsActionsType
$sel:actions:CompromisedCredentialsRiskConfigurationType'

-> CompromisedCredentialsRiskConfigurationType

Create a value of CompromisedCredentialsRiskConfigurationType with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:eventFilter:CompromisedCredentialsRiskConfigurationType', compromisedCredentialsRiskConfigurationType_eventFilter - Perform the action for these events. The default is to perform all events if no event filter is specified.

$sel:actions:CompromisedCredentialsRiskConfigurationType', compromisedCredentialsRiskConfigurationType_actions - The compromised credentials risk configuration actions.

compromisedCredentialsRiskConfigurationType_eventFilter :: Lens' CompromisedCredentialsRiskConfigurationType (Maybe [EventFilterType])

Perform the action for these events. The default is to perform all events if no event filter is specified.

compromisedCredentialsRiskConfigurationType_actions :: Lens' CompromisedCredentialsRiskConfigurationType CompromisedCredentialsActionsType

The compromised credentials risk configuration actions.