Amazonka.CognitoIdentityProvider.Types.NotifyConfigurationType

Description

Documentation

data NotifyConfigurationType

The notify configuration type.

See: newNotifyConfigurationType smart constructor.

Constructors

NotifyConfigurationType'

Fields

noActionEmail :: Maybe NotifyEmailType

The email template used when a detected risk event is allowed.

from :: Maybe Text

The email address that is sending the email. It must be either individually verified with Amazon SES, or from a domain that has been verified with Amazon SES.

replyTo :: Maybe Text

The destination to which the receiver of an email should reply to.

blockEmail :: Maybe NotifyEmailType

Email template used when a detected risk event is blocked.

mfaEmail :: Maybe NotifyEmailType

The MFA email template used when MFA is challenged as part of a detected risk.

sourceArn :: Text

The Amazon Resource Name (ARN) of the identity that is associated with the sending authorization policy. It permits Amazon Cognito to send for the email address specified in the From parameter.

newNotifyConfigurationType

Arguments

:: Text
$sel:sourceArn:NotifyConfigurationType'

-> NotifyConfigurationType

Create a value of NotifyConfigurationType with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:noActionEmail:NotifyConfigurationType', notifyConfigurationType_noActionEmail - The email template used when a detected risk event is allowed.

$sel:from:NotifyConfigurationType', notifyConfigurationType_from - The email address that is sending the email. It must be either individually verified with Amazon SES, or from a domain that has been verified with Amazon SES.

$sel:replyTo:NotifyConfigurationType', notifyConfigurationType_replyTo - The destination to which the receiver of an email should reply to.

$sel:blockEmail:NotifyConfigurationType', notifyConfigurationType_blockEmail - Email template used when a detected risk event is blocked.

$sel:mfaEmail:NotifyConfigurationType', notifyConfigurationType_mfaEmail - The MFA email template used when MFA is challenged as part of a detected risk.

$sel:sourceArn:NotifyConfigurationType', notifyConfigurationType_sourceArn - The Amazon Resource Name (ARN) of the identity that is associated with the sending authorization policy. It permits Amazon Cognito to send for the email address specified in the From parameter.

notifyConfigurationType_noActionEmail :: Lens' NotifyConfigurationType (Maybe NotifyEmailType)

The email template used when a detected risk event is allowed.

notifyConfigurationType_from :: Lens' NotifyConfigurationType (Maybe Text)

The email address that is sending the email. It must be either individually verified with Amazon SES, or from a domain that has been verified with Amazon SES.

notifyConfigurationType_replyTo :: Lens' NotifyConfigurationType (Maybe Text)

The destination to which the receiver of an email should reply to.

notifyConfigurationType_blockEmail :: Lens' NotifyConfigurationType (Maybe NotifyEmailType)

Email template used when a detected risk event is blocked.

notifyConfigurationType_mfaEmail :: Lens' NotifyConfigurationType (Maybe NotifyEmailType)

The MFA email template used when MFA is challenged as part of a detected risk.

notifyConfigurationType_sourceArn :: Lens' NotifyConfigurationType Text

The Amazon Resource Name (ARN) of the identity that is associated with the sending authorization policy. It permits Amazon Cognito to send for the email address specified in the From parameter.