Amazonka.CognitoIdentityProvider.Types.UserImportJobType

Description

Documentation

data UserImportJobType

The user import job type.

See: newUserImportJobType smart constructor.

Constructors

UserImportJobType'

Fields

status :: Maybe UserImportJobStatusType

The status of the user import job. One of the following:

Created - The job was created but not started.

Pending - A transition state. You have started the job, but it has not begun importing users yet.

InProgress - The job has started, and users are being imported.

Stopping - You have stopped the job, but the job has not stopped importing users yet.

Stopped - You have stopped the job, and the job has stopped importing users.

Succeeded - The job has completed successfully.

Failed - The job has stopped due to an error.

Expired - You created a job, but did not start the job within 24-48 hours. All data associated with the job was deleted, and the job cannot be started.

skippedUsers :: Maybe Integer

The number of users that were skipped.

jobId :: Maybe Text

The job ID for the user import job.

userPoolId :: Maybe Text

The user pool ID for the user pool that the users are being imported into.

jobName :: Maybe Text

The job name for the user import job.

preSignedUrl :: Maybe Text

The pre-signed URL to be used to upload the .csv file.

failedUsers :: Maybe Integer

The number of users that could not be imported.

startDate :: Maybe POSIX

The date when the user import job was started.

completionMessage :: Maybe Text

The message returned when the user import job is completed.

creationDate :: Maybe POSIX

The date the user import job was created.

completionDate :: Maybe POSIX

The date when the user import job was completed.

cloudWatchLogsRoleArn :: Maybe Text

The role ARN for the Amazon CloudWatch Logging role for the user import job. For more information, see "Creating the CloudWatch Logs IAM Role" in the Amazon Cognito Developer Guide.

importedUsers :: Maybe Integer

The number of users that were successfully imported.

newUserImportJobType :: UserImportJobType

Create a value of UserImportJobType with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:status:UserImportJobType', userImportJobType_status - The status of the user import job. One of the following:

Created - The job was created but not started.

Pending - A transition state. You have started the job, but it has not begun importing users yet.

InProgress - The job has started, and users are being imported.

Stopping - You have stopped the job, but the job has not stopped importing users yet.

Stopped - You have stopped the job, and the job has stopped importing users.

Succeeded - The job has completed successfully.

Failed - The job has stopped due to an error.

Expired - You created a job, but did not start the job within 24-48 hours. All data associated with the job was deleted, and the job cannot be started.

$sel:skippedUsers:UserImportJobType', userImportJobType_skippedUsers - The number of users that were skipped.

$sel:jobId:UserImportJobType', userImportJobType_jobId - The job ID for the user import job.

$sel:userPoolId:UserImportJobType', userImportJobType_userPoolId - The user pool ID for the user pool that the users are being imported into.

$sel:jobName:UserImportJobType', userImportJobType_jobName - The job name for the user import job.

$sel:preSignedUrl:UserImportJobType', userImportJobType_preSignedUrl - The pre-signed URL to be used to upload the .csv file.

$sel:failedUsers:UserImportJobType', userImportJobType_failedUsers - The number of users that could not be imported.

$sel:startDate:UserImportJobType', userImportJobType_startDate - The date when the user import job was started.

$sel:completionMessage:UserImportJobType', userImportJobType_completionMessage - The message returned when the user import job is completed.

$sel:creationDate:UserImportJobType', userImportJobType_creationDate - The date the user import job was created.

$sel:completionDate:UserImportJobType', userImportJobType_completionDate - The date when the user import job was completed.

$sel:cloudWatchLogsRoleArn:UserImportJobType', userImportJobType_cloudWatchLogsRoleArn - The role ARN for the Amazon CloudWatch Logging role for the user import job. For more information, see "Creating the CloudWatch Logs IAM Role" in the Amazon Cognito Developer Guide.

$sel:importedUsers:UserImportJobType', userImportJobType_importedUsers - The number of users that were successfully imported.

userImportJobType_status :: Lens' UserImportJobType (Maybe UserImportJobStatusType)

The status of the user import job. One of the following:

Created - The job was created but not started.

Pending - A transition state. You have started the job, but it has not begun importing users yet.

InProgress - The job has started, and users are being imported.

Stopping - You have stopped the job, but the job has not stopped importing users yet.

Stopped - You have stopped the job, and the job has stopped importing users.

Succeeded - The job has completed successfully.

Failed - The job has stopped due to an error.

Expired - You created a job, but did not start the job within 24-48 hours. All data associated with the job was deleted, and the job cannot be started.

userImportJobType_skippedUsers :: Lens' UserImportJobType (Maybe Integer)

The number of users that were skipped.

userImportJobType_jobId :: Lens' UserImportJobType (Maybe Text)

The job ID for the user import job.

userImportJobType_userPoolId :: Lens' UserImportJobType (Maybe Text)

The user pool ID for the user pool that the users are being imported into.

userImportJobType_jobName :: Lens' UserImportJobType (Maybe Text)

The job name for the user import job.

userImportJobType_preSignedUrl :: Lens' UserImportJobType (Maybe Text)

The pre-signed URL to be used to upload the .csv file.

userImportJobType_failedUsers :: Lens' UserImportJobType (Maybe Integer)

The number of users that could not be imported.

userImportJobType_startDate :: Lens' UserImportJobType (Maybe UTCTime)

The date when the user import job was started.

userImportJobType_completionMessage :: Lens' UserImportJobType (Maybe Text)

The message returned when the user import job is completed.

userImportJobType_creationDate :: Lens' UserImportJobType (Maybe UTCTime)

The date the user import job was created.

userImportJobType_completionDate :: Lens' UserImportJobType (Maybe UTCTime)

The date when the user import job was completed.

userImportJobType_cloudWatchLogsRoleArn :: Lens' UserImportJobType (Maybe Text)

The role ARN for the Amazon CloudWatch Logging role for the user import job. For more information, see "Creating the CloudWatch Logs IAM Role" in the Amazon Cognito Developer Guide.

userImportJobType_importedUsers :: Lens' UserImportJobType (Maybe Integer)

The number of users that were successfully imported.