Amazonka.CognitoIdentityProvider.UpdateUserAttributes

Description

Allows a user to update a specific attribute (one at a time).

This action might generate an SMS text message. Starting June 1, 2021, U.S. telecom carriers require that you register an origination phone number before you can send SMS messages to U.S. phone numbers. If you use SMS text messages in Amazon Cognito, you must register a phone number with Amazon Pinpoint. Cognito will use the the registered number automatically. Otherwise, Cognito users that must receive SMS messages might be unable to sign up, activate their accounts, or sign in.

If you have never used SMS text messages with Amazon Cognito or any other Amazon Web Service, Amazon SNS might place your account in SMS sandbox. In <https:docs.aws.amazon.comsnslatestdgsns-sms-sandbox.html sandbox mode> , you’ll have limitations, such as sending messages to only verified phone numbers. After testing in the sandbox environment, you can move out of the SMS sandbox and into production. For more information, see SMS message settings for Cognito User Pools in the Amazon Cognito Developer Guide.

Creating a Request

data UpdateUserAttributes

Represents the request to update user attributes.

See: newUpdateUserAttributes smart constructor.

Constructors

UpdateUserAttributes'

Fields

clientMetadata :: Maybe (HashMap Text Text)

A map of custom key-value pairs that you can provide as input for any custom workflows that this action triggers.

You create custom workflows by assigning Lambda functions to user pool triggers. When you use the UpdateUserAttributes API action, Amazon Cognito invokes the function that is assigned to the custom message trigger. When Amazon Cognito invokes this function, it passes a JSON payload, which the function receives as input. This payload contains a clientMetadata attribute, which provides the data that you assigned to the ClientMetadata parameter in your UpdateUserAttributes request. In your function code in Lambda, you can process the clientMetadata value to enhance your workflow for your specific needs.

For more information, see Customizing User Pool Workflows with Lambda Triggers in the Amazon Cognito Developer Guide.

Take the following limitations into consideration when you use the ClientMetadata parameter:

Amazon Cognito does not store the ClientMetadata value. This data is available only to Lambda triggers that are assigned to a user pool to support custom workflows. If your user pool configuration does not include triggers, the ClientMetadata parameter serves no purpose.

Amazon Cognito does not validate the ClientMetadata value.

Amazon Cognito does not encrypt the the ClientMetadata value, so don't use it to provide sensitive information.

userAttributes :: [AttributeType]

An array of name-value pairs representing user attributes.

For custom attributes, you must prepend the custom: prefix to the attribute name.

accessToken :: Sensitive Text

The access token for the request to update user attributes.

newUpdateUserAttributes

Arguments

:: Text
$sel:accessToken:UpdateUserAttributes'

-> UpdateUserAttributes

Create a value of UpdateUserAttributes with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:clientMetadata:UpdateUserAttributes', updateUserAttributes_clientMetadata - A map of custom key-value pairs that you can provide as input for any custom workflows that this action triggers.

You create custom workflows by assigning Lambda functions to user pool triggers. When you use the UpdateUserAttributes API action, Amazon Cognito invokes the function that is assigned to the custom message trigger. When Amazon Cognito invokes this function, it passes a JSON payload, which the function receives as input. This payload contains a clientMetadata attribute, which provides the data that you assigned to the ClientMetadata parameter in your UpdateUserAttributes request. In your function code in Lambda, you can process the clientMetadata value to enhance your workflow for your specific needs.

For more information, see Customizing User Pool Workflows with Lambda Triggers in the Amazon Cognito Developer Guide.

Take the following limitations into consideration when you use the ClientMetadata parameter:

Amazon Cognito does not store the ClientMetadata value. This data is available only to Lambda triggers that are assigned to a user pool to support custom workflows. If your user pool configuration does not include triggers, the ClientMetadata parameter serves no purpose.

Amazon Cognito does not validate the ClientMetadata value.

Amazon Cognito does not encrypt the the ClientMetadata value, so don't use it to provide sensitive information.

$sel:userAttributes:UpdateUserAttributes', updateUserAttributes_userAttributes - An array of name-value pairs representing user attributes.

For custom attributes, you must prepend the custom: prefix to the attribute name.

$sel:accessToken:UpdateUserAttributes', updateUserAttributes_accessToken - The access token for the request to update user attributes.

Request Lenses

updateUserAttributes_clientMetadata :: Lens' UpdateUserAttributes (Maybe (HashMap Text Text))

A map of custom key-value pairs that you can provide as input for any custom workflows that this action triggers.

You create custom workflows by assigning Lambda functions to user pool triggers. When you use the UpdateUserAttributes API action, Amazon Cognito invokes the function that is assigned to the custom message trigger. When Amazon Cognito invokes this function, it passes a JSON payload, which the function receives as input. This payload contains a clientMetadata attribute, which provides the data that you assigned to the ClientMetadata parameter in your UpdateUserAttributes request. In your function code in Lambda, you can process the clientMetadata value to enhance your workflow for your specific needs.

For more information, see Customizing User Pool Workflows with Lambda Triggers in the Amazon Cognito Developer Guide.

Take the following limitations into consideration when you use the ClientMetadata parameter:

Amazon Cognito does not store the ClientMetadata value. This data is available only to Lambda triggers that are assigned to a user pool to support custom workflows. If your user pool configuration does not include triggers, the ClientMetadata parameter serves no purpose.

Amazon Cognito does not validate the ClientMetadata value.

Amazon Cognito does not encrypt the the ClientMetadata value, so don't use it to provide sensitive information.

updateUserAttributes_userAttributes :: Lens' UpdateUserAttributes [AttributeType]

An array of name-value pairs representing user attributes.

For custom attributes, you must prepend the custom: prefix to the attribute name.

updateUserAttributes_accessToken :: Lens' UpdateUserAttributes Text

The access token for the request to update user attributes.

Destructuring the Response

data UpdateUserAttributesResponse

Represents the response from the server for the request to update user attributes.

See: newUpdateUserAttributesResponse smart constructor.

Constructors

UpdateUserAttributesResponse'

Fields

codeDeliveryDetailsList :: Maybe [CodeDeliveryDetailsType]

The code delivery details list from the server for the request to update user attributes.

httpStatus :: Int

The response's http status code.

newUpdateUserAttributesResponse

Arguments

:: Int
$sel:httpStatus:UpdateUserAttributesResponse'

-> UpdateUserAttributesResponse

Create a value of UpdateUserAttributesResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:codeDeliveryDetailsList:UpdateUserAttributesResponse', updateUserAttributesResponse_codeDeliveryDetailsList - The code delivery details list from the server for the request to update user attributes.

$sel:httpStatus:UpdateUserAttributesResponse', updateUserAttributesResponse_httpStatus - The response's http status code.

Response Lenses

updateUserAttributesResponse_codeDeliveryDetailsList :: Lens' UpdateUserAttributesResponse (Maybe [CodeDeliveryDetailsType])

The code delivery details list from the server for the request to update user attributes.

updateUserAttributesResponse_httpStatus :: Lens' UpdateUserAttributesResponse Int

The response's http status code.