Amazonka.CognitoIdentityProvider.UpdateUserPool

Description

Updates the specified user pool with the specified attributes. You can get a list of the current user pool settings using DescribeUserPool. If you don't provide a value for an attribute, it will be set to the default value.

This action might generate an SMS text message. Starting June 1, 2021, U.S. telecom carriers require that you register an origination phone number before you can send SMS messages to U.S. phone numbers. If you use SMS text messages in Amazon Cognito, you must register a phone number with Amazon Pinpoint. Cognito will use the the registered number automatically. Otherwise, Cognito users that must receive SMS messages might be unable to sign up, activate their accounts, or sign in.

If you have never used SMS text messages with Amazon Cognito or any other Amazon Web Service, Amazon SNS might place your account in SMS sandbox. In <https:docs.aws.amazon.comsnslatestdgsns-sms-sandbox.html sandbox mode> , you’ll have limitations, such as sending messages to only verified phone numbers. After testing in the sandbox environment, you can move out of the SMS sandbox and into production. For more information, see SMS message settings for Cognito User Pools in the Amazon Cognito Developer Guide.

Creating a Request

data UpdateUserPool

Represents the request to update the user pool.

See: newUpdateUserPool smart constructor.

Constructors

UpdateUserPool'

Fields

userPoolTags :: Maybe (HashMap Text Text)

The tag keys and values to assign to the user pool. A tag is a label that you can use to categorize and manage user pools in different ways, such as by purpose, owner, environment, or other criteria.

verificationMessageTemplate :: Maybe VerificationMessageTemplateType

The template for verification messages.

emailVerificationMessage :: Maybe Text

The contents of the email verification message.

smsAuthenticationMessage :: Maybe Text

The contents of the SMS authentication message.

userPoolAddOns :: Maybe UserPoolAddOnsType

Used to enable advanced security risk detection. Set the key AdvancedSecurityMode to the value "AUDIT".

emailVerificationSubject :: Maybe Text

The subject of the email verification message.

accountRecoverySetting :: Maybe AccountRecoverySettingType

Use this setting to define which verified available method a user can use to recover their password when they call ForgotPassword. It allows you to define a preferred method when a user has more than one method available. With this setting, SMS does not qualify for a valid password recovery mechanism if the user also has SMS MFA enabled. In the absence of this setting, Cognito uses the legacy behavior to determine the recovery method where SMS is preferred over email.

emailConfiguration :: Maybe EmailConfigurationType

Email configuration.

smsVerificationMessage :: Maybe Text

A container with information about the SMS verification message.

mfaConfiguration :: Maybe UserPoolMfaType

Can be one of the following values:

OFF - MFA tokens are not required and cannot be specified during user registration.

ON - MFA tokens are required for all user registrations. You can only specify ON when you are initially creating a user pool. You can use the SetUserPoolMfaConfig API operation to turn MFA "ON" for existing user pools.

OPTIONAL - Users have the option when registering to create an MFA token.

lambdaConfig :: Maybe LambdaConfigType

The Lambda configuration information from the request to update the user pool.

smsConfiguration :: Maybe SmsConfigurationType

SMS configuration.

adminCreateUserConfig :: Maybe AdminCreateUserConfigType

The configuration for AdminCreateUser requests.

deviceConfiguration :: Maybe DeviceConfigurationType

Device configuration.

autoVerifiedAttributes :: Maybe [VerifiedAttributeType]

The attributes that are automatically verified when the Amazon Cognito service makes a request to update user pools.

policies :: Maybe UserPoolPolicyType

A container with the policies you wish to update in a user pool.

userPoolId :: Text

The user pool ID for the user pool you want to update.

newUpdateUserPool

Arguments

:: Text
$sel:userPoolId:UpdateUserPool'

-> UpdateUserPool

Create a value of UpdateUserPool with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:userPoolTags:UpdateUserPool', updateUserPool_userPoolTags - The tag keys and values to assign to the user pool. A tag is a label that you can use to categorize and manage user pools in different ways, such as by purpose, owner, environment, or other criteria.

$sel:verificationMessageTemplate:UpdateUserPool', updateUserPool_verificationMessageTemplate - The template for verification messages.

$sel:emailVerificationMessage:UpdateUserPool', updateUserPool_emailVerificationMessage - The contents of the email verification message.

$sel:smsAuthenticationMessage:UpdateUserPool', updateUserPool_smsAuthenticationMessage - The contents of the SMS authentication message.

$sel:userPoolAddOns:UpdateUserPool', updateUserPool_userPoolAddOns - Used to enable advanced security risk detection. Set the key AdvancedSecurityMode to the value "AUDIT".

$sel:emailVerificationSubject:UpdateUserPool', updateUserPool_emailVerificationSubject - The subject of the email verification message.

$sel:accountRecoverySetting:UpdateUserPool', updateUserPool_accountRecoverySetting - Use this setting to define which verified available method a user can use to recover their password when they call ForgotPassword. It allows you to define a preferred method when a user has more than one method available. With this setting, SMS does not qualify for a valid password recovery mechanism if the user also has SMS MFA enabled. In the absence of this setting, Cognito uses the legacy behavior to determine the recovery method where SMS is preferred over email.

$sel:emailConfiguration:UpdateUserPool', updateUserPool_emailConfiguration - Email configuration.

$sel:smsVerificationMessage:UpdateUserPool', updateUserPool_smsVerificationMessage - A container with information about the SMS verification message.

$sel:mfaConfiguration:UpdateUserPool', updateUserPool_mfaConfiguration - Can be one of the following values:

OFF - MFA tokens are not required and cannot be specified during user registration.

ON - MFA tokens are required for all user registrations. You can only specify ON when you are initially creating a user pool. You can use the SetUserPoolMfaConfig API operation to turn MFA "ON" for existing user pools.

OPTIONAL - Users have the option when registering to create an MFA token.

$sel:lambdaConfig:UpdateUserPool', updateUserPool_lambdaConfig - The Lambda configuration information from the request to update the user pool.

$sel:smsConfiguration:UpdateUserPool', updateUserPool_smsConfiguration - SMS configuration.

$sel:adminCreateUserConfig:UpdateUserPool', updateUserPool_adminCreateUserConfig - The configuration for AdminCreateUser requests.

$sel:deviceConfiguration:UpdateUserPool', updateUserPool_deviceConfiguration - Device configuration.

$sel:autoVerifiedAttributes:UpdateUserPool', updateUserPool_autoVerifiedAttributes - The attributes that are automatically verified when the Amazon Cognito service makes a request to update user pools.

$sel:policies:UpdateUserPool', updateUserPool_policies - A container with the policies you wish to update in a user pool.

$sel:userPoolId:UpdateUserPool', updateUserPool_userPoolId - The user pool ID for the user pool you want to update.

Request Lenses

updateUserPool_userPoolTags :: Lens' UpdateUserPool (Maybe (HashMap Text Text))

The tag keys and values to assign to the user pool. A tag is a label that you can use to categorize and manage user pools in different ways, such as by purpose, owner, environment, or other criteria.

updateUserPool_verificationMessageTemplate :: Lens' UpdateUserPool (Maybe VerificationMessageTemplateType)

The template for verification messages.

updateUserPool_emailVerificationMessage :: Lens' UpdateUserPool (Maybe Text)

The contents of the email verification message.

updateUserPool_smsAuthenticationMessage :: Lens' UpdateUserPool (Maybe Text)

The contents of the SMS authentication message.

updateUserPool_userPoolAddOns :: Lens' UpdateUserPool (Maybe UserPoolAddOnsType)

Used to enable advanced security risk detection. Set the key AdvancedSecurityMode to the value "AUDIT".

updateUserPool_emailVerificationSubject :: Lens' UpdateUserPool (Maybe Text)

The subject of the email verification message.

updateUserPool_accountRecoverySetting :: Lens' UpdateUserPool (Maybe AccountRecoverySettingType)

Use this setting to define which verified available method a user can use to recover their password when they call ForgotPassword. It allows you to define a preferred method when a user has more than one method available. With this setting, SMS does not qualify for a valid password recovery mechanism if the user also has SMS MFA enabled. In the absence of this setting, Cognito uses the legacy behavior to determine the recovery method where SMS is preferred over email.

updateUserPool_emailConfiguration :: Lens' UpdateUserPool (Maybe EmailConfigurationType)

Email configuration.

updateUserPool_smsVerificationMessage :: Lens' UpdateUserPool (Maybe Text)

A container with information about the SMS verification message.

updateUserPool_mfaConfiguration :: Lens' UpdateUserPool (Maybe UserPoolMfaType)

Can be one of the following values:

OFF - MFA tokens are not required and cannot be specified during user registration.

ON - MFA tokens are required for all user registrations. You can only specify ON when you are initially creating a user pool. You can use the SetUserPoolMfaConfig API operation to turn MFA "ON" for existing user pools.

OPTIONAL - Users have the option when registering to create an MFA token.

updateUserPool_lambdaConfig :: Lens' UpdateUserPool (Maybe LambdaConfigType)

The Lambda configuration information from the request to update the user pool.

updateUserPool_smsConfiguration :: Lens' UpdateUserPool (Maybe SmsConfigurationType)

SMS configuration.

updateUserPool_adminCreateUserConfig :: Lens' UpdateUserPool (Maybe AdminCreateUserConfigType)

The configuration for AdminCreateUser requests.

updateUserPool_deviceConfiguration :: Lens' UpdateUserPool (Maybe DeviceConfigurationType)

Device configuration.

updateUserPool_autoVerifiedAttributes :: Lens' UpdateUserPool (Maybe [VerifiedAttributeType])

The attributes that are automatically verified when the Amazon Cognito service makes a request to update user pools.

updateUserPool_policies :: Lens' UpdateUserPool (Maybe UserPoolPolicyType)

A container with the policies you wish to update in a user pool.

updateUserPool_userPoolId :: Lens' UpdateUserPool Text

The user pool ID for the user pool you want to update.

Destructuring the Response

data UpdateUserPoolResponse

Represents the response from the server when you make a request to update the user pool.

See: newUpdateUserPoolResponse smart constructor.

Constructors

UpdateUserPoolResponse'

Fields

httpStatus :: Int

The response's http status code.

newUpdateUserPoolResponse

Arguments

:: Int
$sel:httpStatus:UpdateUserPoolResponse'

-> UpdateUserPoolResponse

Create a value of UpdateUserPoolResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:httpStatus:UpdateUserPoolResponse', updateUserPoolResponse_httpStatus - The response's http status code.

Response Lenses

updateUserPoolResponse_httpStatus :: Lens' UpdateUserPoolResponse Int

The response's http status code.