Amazonka.CognitoSync.Types.Dataset

Description

Documentation

data Dataset

A collection of data for an identity pool. An identity pool can have multiple datasets. A dataset is per identity and can be general or associated with a particular entity in an application (like a saved game). Datasets are automatically created if they don't exist. Data is synced by dataset, and a dataset can hold up to 1MB of key-value pairs.

See: newDataset smart constructor.

Constructors

Dataset'

Fields

lastModifiedDate :: Maybe POSIX

Date when the dataset was last modified.

numRecords :: Maybe Integer

Number of records in this dataset.

dataStorage :: Maybe Integer

Total size in bytes of the records in this dataset.

datasetName :: Maybe Text

A string of up to 128 characters. Allowed characters are a-z, A-Z, 0-9, '_' (underscore), '-' (dash), and '.' (dot).

creationDate :: Maybe POSIX

Date on which the dataset was created.

lastModifiedBy :: Maybe Text

The device that made the last change to this dataset.

identityId :: Maybe Text

A name-spaced GUID (for example, us-east-1:23EC4050-6AEA-7089-A2DD-08002EXAMPLE) created by Amazon Cognito. GUID generation is unique within a region.

newDataset :: Dataset

Create a value of Dataset with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:lastModifiedDate:Dataset', dataset_lastModifiedDate - Date when the dataset was last modified.

$sel:numRecords:Dataset', dataset_numRecords - Number of records in this dataset.

$sel:dataStorage:Dataset', dataset_dataStorage - Total size in bytes of the records in this dataset.

$sel:datasetName:Dataset', dataset_datasetName - A string of up to 128 characters. Allowed characters are a-z, A-Z, 0-9, '_' (underscore), '-' (dash), and '.' (dot).

$sel:creationDate:Dataset', dataset_creationDate - Date on which the dataset was created.

$sel:lastModifiedBy:Dataset', dataset_lastModifiedBy - The device that made the last change to this dataset.

$sel:identityId:Dataset', dataset_identityId - A name-spaced GUID (for example, us-east-1:23EC4050-6AEA-7089-A2DD-08002EXAMPLE) created by Amazon Cognito. GUID generation is unique within a region.

dataset_lastModifiedDate :: Lens' Dataset (Maybe UTCTime)

Date when the dataset was last modified.

dataset_numRecords :: Lens' Dataset (Maybe Integer)

Number of records in this dataset.

dataset_dataStorage :: Lens' Dataset (Maybe Integer)

Total size in bytes of the records in this dataset.

dataset_datasetName :: Lens' Dataset (Maybe Text)

A string of up to 128 characters. Allowed characters are a-z, A-Z, 0-9, '_' (underscore), '-' (dash), and '.' (dot).

dataset_creationDate :: Lens' Dataset (Maybe UTCTime)

Date on which the dataset was created.

dataset_lastModifiedBy :: Lens' Dataset (Maybe Text)

The device that made the last change to this dataset.

dataset_identityId :: Lens' Dataset (Maybe Text)

A name-spaced GUID (for example, us-east-1:23EC4050-6AEA-7089-A2DD-08002EXAMPLE) created by Amazon Cognito. GUID generation is unique within a region.