Amazonka.Comprehend.BatchDetectEntities

Description

Inspects the text of a batch of documents for named entities and returns information about them. For more information about named entities, see how-entities

Creating a Request

data BatchDetectEntities

See: newBatchDetectEntities smart constructor.

Constructors

BatchDetectEntities'

Fields

textList :: Sensitive [Sensitive Text]

A list containing the text of the input documents. The list can contain a maximum of 25 documents. Each document must contain fewer than 5,000 bytes of UTF-8 encoded characters.

languageCode :: LanguageCode

The language of the input documents. You can specify any of the primary languages supported by Amazon Comprehend. All documents must be in the same language.

newBatchDetectEntities

Arguments

:: LanguageCode
$sel:languageCode:BatchDetectEntities'

-> BatchDetectEntities

Create a value of BatchDetectEntities with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:textList:BatchDetectEntities', batchDetectEntities_textList - A list containing the text of the input documents. The list can contain a maximum of 25 documents. Each document must contain fewer than 5,000 bytes of UTF-8 encoded characters.

$sel:languageCode:BatchDetectEntities', batchDetectEntities_languageCode - The language of the input documents. You can specify any of the primary languages supported by Amazon Comprehend. All documents must be in the same language.

Request Lenses

batchDetectEntities_textList :: Lens' BatchDetectEntities [Text]

A list containing the text of the input documents. The list can contain a maximum of 25 documents. Each document must contain fewer than 5,000 bytes of UTF-8 encoded characters.

batchDetectEntities_languageCode :: Lens' BatchDetectEntities LanguageCode

The language of the input documents. You can specify any of the primary languages supported by Amazon Comprehend. All documents must be in the same language.

Destructuring the Response

data BatchDetectEntitiesResponse

See: newBatchDetectEntitiesResponse smart constructor.

Constructors

BatchDetectEntitiesResponse'

Fields

httpStatus :: Int

The response's http status code.

resultList :: [BatchDetectEntitiesItemResult]

A list of objects containing the results of the operation. The results are sorted in ascending order by the Index field and match the order of the documents in the input list. If all of the documents contain an error, the ResultList is empty.

errorList :: [BatchItemError]

A list containing one object for each document that contained an error. The results are sorted in ascending order by the Index field and match the order of the documents in the input list. If there are no errors in the batch, the ErrorList is empty.

newBatchDetectEntitiesResponse

Arguments

:: Int
$sel:httpStatus:BatchDetectEntitiesResponse'

-> BatchDetectEntitiesResponse

Create a value of BatchDetectEntitiesResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:httpStatus:BatchDetectEntitiesResponse', batchDetectEntitiesResponse_httpStatus - The response's http status code.

$sel:resultList:BatchDetectEntitiesResponse', batchDetectEntitiesResponse_resultList - A list of objects containing the results of the operation. The results are sorted in ascending order by the Index field and match the order of the documents in the input list. If all of the documents contain an error, the ResultList is empty.

$sel:errorList:BatchDetectEntitiesResponse', batchDetectEntitiesResponse_errorList - A list containing one object for each document that contained an error. The results are sorted in ascending order by the Index field and match the order of the documents in the input list. If there are no errors in the batch, the ErrorList is empty.

Response Lenses

batchDetectEntitiesResponse_httpStatus :: Lens' BatchDetectEntitiesResponse Int

The response's http status code.

batchDetectEntitiesResponse_resultList :: Lens' BatchDetectEntitiesResponse [BatchDetectEntitiesItemResult]

A list of objects containing the results of the operation. The results are sorted in ascending order by the Index field and match the order of the documents in the input list. If all of the documents contain an error, the ResultList is empty.

batchDetectEntitiesResponse_errorList :: Lens' BatchDetectEntitiesResponse [BatchItemError]

A list containing one object for each document that contained an error. The results are sorted in ascending order by the Index field and match the order of the documents in the input list. If there are no errors in the batch, the ErrorList is empty.