Amazonka.Comprehend.BatchDetectKeyPhrases

Description

Detects the key noun phrases found in a batch of documents.

Creating a Request

data BatchDetectKeyPhrases

See: newBatchDetectKeyPhrases smart constructor.

Constructors

BatchDetectKeyPhrases'

Fields

textList :: Sensitive [Sensitive Text]

A list containing the text of the input documents. The list can contain a maximum of 25 documents. Each document must contain fewer that 5,000 bytes of UTF-8 encoded characters.

languageCode :: LanguageCode

The language of the input documents. You can specify any of the primary languages supported by Amazon Comprehend. All documents must be in the same language.

newBatchDetectKeyPhrases

Arguments

:: LanguageCode
$sel:languageCode:BatchDetectKeyPhrases'

-> BatchDetectKeyPhrases

Create a value of BatchDetectKeyPhrases with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:textList:BatchDetectKeyPhrases', batchDetectKeyPhrases_textList - A list containing the text of the input documents. The list can contain a maximum of 25 documents. Each document must contain fewer that 5,000 bytes of UTF-8 encoded characters.

$sel:languageCode:BatchDetectKeyPhrases', batchDetectKeyPhrases_languageCode - The language of the input documents. You can specify any of the primary languages supported by Amazon Comprehend. All documents must be in the same language.

Request Lenses

batchDetectKeyPhrases_textList :: Lens' BatchDetectKeyPhrases [Text]

A list containing the text of the input documents. The list can contain a maximum of 25 documents. Each document must contain fewer that 5,000 bytes of UTF-8 encoded characters.

batchDetectKeyPhrases_languageCode :: Lens' BatchDetectKeyPhrases LanguageCode

The language of the input documents. You can specify any of the primary languages supported by Amazon Comprehend. All documents must be in the same language.

Destructuring the Response

data BatchDetectKeyPhrasesResponse

See: newBatchDetectKeyPhrasesResponse smart constructor.

Constructors

BatchDetectKeyPhrasesResponse'

Fields

httpStatus :: Int

The response's http status code.

resultList :: [BatchDetectKeyPhrasesItemResult]

A list of objects containing the results of the operation. The results are sorted in ascending order by the Index field and match the order of the documents in the input list. If all of the documents contain an error, the ResultList is empty.

errorList :: [BatchItemError]

A list containing one object for each document that contained an error. The results are sorted in ascending order by the Index field and match the order of the documents in the input list. If there are no errors in the batch, the ErrorList is empty.

newBatchDetectKeyPhrasesResponse

Arguments

:: Int
$sel:httpStatus:BatchDetectKeyPhrasesResponse'

-> BatchDetectKeyPhrasesResponse

Create a value of BatchDetectKeyPhrasesResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:httpStatus:BatchDetectKeyPhrasesResponse', batchDetectKeyPhrasesResponse_httpStatus - The response's http status code.

$sel:resultList:BatchDetectKeyPhrasesResponse', batchDetectKeyPhrasesResponse_resultList - A list of objects containing the results of the operation. The results are sorted in ascending order by the Index field and match the order of the documents in the input list. If all of the documents contain an error, the ResultList is empty.

$sel:errorList:BatchDetectKeyPhrasesResponse', batchDetectKeyPhrasesResponse_errorList - A list containing one object for each document that contained an error. The results are sorted in ascending order by the Index field and match the order of the documents in the input list. If there are no errors in the batch, the ErrorList is empty.

Response Lenses

batchDetectKeyPhrasesResponse_httpStatus :: Lens' BatchDetectKeyPhrasesResponse Int

The response's http status code.

batchDetectKeyPhrasesResponse_resultList :: Lens' BatchDetectKeyPhrasesResponse [BatchDetectKeyPhrasesItemResult]

A list of objects containing the results of the operation. The results are sorted in ascending order by the Index field and match the order of the documents in the input list. If all of the documents contain an error, the ResultList is empty.

batchDetectKeyPhrasesResponse_errorList :: Lens' BatchDetectKeyPhrasesResponse [BatchItemError]

A list containing one object for each document that contained an error. The results are sorted in ascending order by the Index field and match the order of the documents in the input list. If there are no errors in the batch, the ErrorList is empty.