Amazonka.Comprehend.CreateEntityRecognizer

Description

Creates an entity recognizer using submitted files. After your CreateEntityRecognizer request is submitted, you can check job status using the API.

Creating a Request

data CreateEntityRecognizer

See: newCreateEntityRecognizer smart constructor.

Constructors

CreateEntityRecognizer'

Fields

versionName :: Maybe Text

The version name given to the newly created recognizer. Version names can be a maximum of 256 characters. Alphanumeric characters, hyphens (-) and underscores (_) are allowed. The version name must be unique among all models with the same recognizer name in the account/ AWS Region.

modelKmsKeyId :: Maybe Text

ID for the AWS Key Management Service (KMS) key that Amazon Comprehend uses to encrypt trained custom models. The ModelKmsKeyId can be either of the following formats

KMS Key ID: "1234abcd-12ab-34cd-56ef-1234567890ab"

Amazon Resource Name (ARN) of a KMS Key: "arn:aws:kms:us-west-2:111122223333:key/1234abcd-12ab-34cd-56ef-1234567890ab"

vpcConfig :: Maybe VpcConfig

Configuration parameters for an optional private Virtual Private Cloud (VPC) containing the resources you are using for your custom entity recognizer. For more information, see Amazon VPC.

volumeKmsKeyId :: Maybe Text

ID for the AWS Key Management Service (KMS) key that Amazon Comprehend uses to encrypt data on the storage volume attached to the ML compute instance(s) that process the analysis job. The VolumeKmsKeyId can be either of the following formats:

KMS Key ID: "1234abcd-12ab-34cd-56ef-1234567890ab"

Amazon Resource Name (ARN) of a KMS Key: "arn:aws:kms:us-west-2:111122223333:key/1234abcd-12ab-34cd-56ef-1234567890ab"

clientRequestToken :: Maybe Text

A unique identifier for the request. If you don't set the client request token, Amazon Comprehend generates one.

tags :: Maybe [Tag]

Tags to be associated with the entity recognizer being created. A tag is a key-value pair that adds as a metadata to a resource used by Amazon Comprehend. For example, a tag with "Sales" as the key might be added to a resource to indicate its use by the sales department.

recognizerName :: Text

The name given to the newly created recognizer. Recognizer names can be a maximum of 256 characters. Alphanumeric characters, hyphens (-) and underscores (_) are allowed. The name must be unique in the account/region.

dataAccessRoleArn :: Text

The Amazon Resource Name (ARN) of the AWS Identity and Management (IAM) role that grants Amazon Comprehend read access to your input data.

inputDataConfig :: EntityRecognizerInputDataConfig

Specifies the format and location of the input data. The S3 bucket containing the input data must be located in the same region as the entity recognizer being created.

languageCode :: LanguageCode

You can specify any of the following languages supported by Amazon Comprehend: English ("en"), Spanish ("es"), French ("fr"), Italian ("it"), German ("de"), or Portuguese ("pt"). All documents must be in the same language.

newCreateEntityRecognizer

Arguments

:: Text
$sel:recognizerName:CreateEntityRecognizer'

-> Text
$sel:dataAccessRoleArn:CreateEntityRecognizer'

-> EntityRecognizerInputDataConfig
$sel:inputDataConfig:CreateEntityRecognizer'

-> LanguageCode
$sel:languageCode:CreateEntityRecognizer'

-> CreateEntityRecognizer

Create a value of CreateEntityRecognizer with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:versionName:CreateEntityRecognizer', createEntityRecognizer_versionName - The version name given to the newly created recognizer. Version names can be a maximum of 256 characters. Alphanumeric characters, hyphens (-) and underscores (_) are allowed. The version name must be unique among all models with the same recognizer name in the account/ AWS Region.

$sel:modelKmsKeyId:CreateEntityRecognizer', createEntityRecognizer_modelKmsKeyId - ID for the AWS Key Management Service (KMS) key that Amazon Comprehend uses to encrypt trained custom models. The ModelKmsKeyId can be either of the following formats

KMS Key ID: "1234abcd-12ab-34cd-56ef-1234567890ab"

Amazon Resource Name (ARN) of a KMS Key: "arn:aws:kms:us-west-2:111122223333:key/1234abcd-12ab-34cd-56ef-1234567890ab"

$sel:vpcConfig:CreateEntityRecognizer', createEntityRecognizer_vpcConfig - Configuration parameters for an optional private Virtual Private Cloud (VPC) containing the resources you are using for your custom entity recognizer. For more information, see Amazon VPC.

$sel:volumeKmsKeyId:CreateEntityRecognizer', createEntityRecognizer_volumeKmsKeyId - ID for the AWS Key Management Service (KMS) key that Amazon Comprehend uses to encrypt data on the storage volume attached to the ML compute instance(s) that process the analysis job. The VolumeKmsKeyId can be either of the following formats:

KMS Key ID: "1234abcd-12ab-34cd-56ef-1234567890ab"

Amazon Resource Name (ARN) of a KMS Key: "arn:aws:kms:us-west-2:111122223333:key/1234abcd-12ab-34cd-56ef-1234567890ab"

$sel:clientRequestToken:CreateEntityRecognizer', createEntityRecognizer_clientRequestToken - A unique identifier for the request. If you don't set the client request token, Amazon Comprehend generates one.

$sel:tags:CreateEntityRecognizer', createEntityRecognizer_tags - Tags to be associated with the entity recognizer being created. A tag is a key-value pair that adds as a metadata to a resource used by Amazon Comprehend. For example, a tag with "Sales" as the key might be added to a resource to indicate its use by the sales department.

$sel:recognizerName:CreateEntityRecognizer', createEntityRecognizer_recognizerName - The name given to the newly created recognizer. Recognizer names can be a maximum of 256 characters. Alphanumeric characters, hyphens (-) and underscores (_) are allowed. The name must be unique in the account/region.

$sel:dataAccessRoleArn:CreateEntityRecognizer', createEntityRecognizer_dataAccessRoleArn - The Amazon Resource Name (ARN) of the AWS Identity and Management (IAM) role that grants Amazon Comprehend read access to your input data.

$sel:inputDataConfig:CreateEntityRecognizer', createEntityRecognizer_inputDataConfig - Specifies the format and location of the input data. The S3 bucket containing the input data must be located in the same region as the entity recognizer being created.

$sel:languageCode:CreateEntityRecognizer', createEntityRecognizer_languageCode - You can specify any of the following languages supported by Amazon Comprehend: English ("en"), Spanish ("es"), French ("fr"), Italian ("it"), German ("de"), or Portuguese ("pt"). All documents must be in the same language.

Request Lenses

createEntityRecognizer_versionName :: Lens' CreateEntityRecognizer (Maybe Text)

The version name given to the newly created recognizer. Version names can be a maximum of 256 characters. Alphanumeric characters, hyphens (-) and underscores (_) are allowed. The version name must be unique among all models with the same recognizer name in the account/ AWS Region.

createEntityRecognizer_modelKmsKeyId :: Lens' CreateEntityRecognizer (Maybe Text)

ID for the AWS Key Management Service (KMS) key that Amazon Comprehend uses to encrypt trained custom models. The ModelKmsKeyId can be either of the following formats

KMS Key ID: "1234abcd-12ab-34cd-56ef-1234567890ab"

Amazon Resource Name (ARN) of a KMS Key: "arn:aws:kms:us-west-2:111122223333:key/1234abcd-12ab-34cd-56ef-1234567890ab"

createEntityRecognizer_vpcConfig :: Lens' CreateEntityRecognizer (Maybe VpcConfig)

Configuration parameters for an optional private Virtual Private Cloud (VPC) containing the resources you are using for your custom entity recognizer. For more information, see Amazon VPC.

createEntityRecognizer_volumeKmsKeyId :: Lens' CreateEntityRecognizer (Maybe Text)

ID for the AWS Key Management Service (KMS) key that Amazon Comprehend uses to encrypt data on the storage volume attached to the ML compute instance(s) that process the analysis job. The VolumeKmsKeyId can be either of the following formats:

KMS Key ID: "1234abcd-12ab-34cd-56ef-1234567890ab"

Amazon Resource Name (ARN) of a KMS Key: "arn:aws:kms:us-west-2:111122223333:key/1234abcd-12ab-34cd-56ef-1234567890ab"

createEntityRecognizer_clientRequestToken :: Lens' CreateEntityRecognizer (Maybe Text)

A unique identifier for the request. If you don't set the client request token, Amazon Comprehend generates one.

createEntityRecognizer_tags :: Lens' CreateEntityRecognizer (Maybe [Tag])

Tags to be associated with the entity recognizer being created. A tag is a key-value pair that adds as a metadata to a resource used by Amazon Comprehend. For example, a tag with "Sales" as the key might be added to a resource to indicate its use by the sales department.

createEntityRecognizer_recognizerName :: Lens' CreateEntityRecognizer Text

The name given to the newly created recognizer. Recognizer names can be a maximum of 256 characters. Alphanumeric characters, hyphens (-) and underscores (_) are allowed. The name must be unique in the account/region.

createEntityRecognizer_dataAccessRoleArn :: Lens' CreateEntityRecognizer Text

The Amazon Resource Name (ARN) of the AWS Identity and Management (IAM) role that grants Amazon Comprehend read access to your input data.

createEntityRecognizer_inputDataConfig :: Lens' CreateEntityRecognizer EntityRecognizerInputDataConfig

Specifies the format and location of the input data. The S3 bucket containing the input data must be located in the same region as the entity recognizer being created.

createEntityRecognizer_languageCode :: Lens' CreateEntityRecognizer LanguageCode

You can specify any of the following languages supported by Amazon Comprehend: English ("en"), Spanish ("es"), French ("fr"), Italian ("it"), German ("de"), or Portuguese ("pt"). All documents must be in the same language.

Destructuring the Response

data CreateEntityRecognizerResponse

See: newCreateEntityRecognizerResponse smart constructor.

Constructors

CreateEntityRecognizerResponse'

Fields

entityRecognizerArn :: Maybe Text

The Amazon Resource Name (ARN) that identifies the entity recognizer.

httpStatus :: Int

The response's http status code.

newCreateEntityRecognizerResponse

Arguments

:: Int
$sel:httpStatus:CreateEntityRecognizerResponse'

-> CreateEntityRecognizerResponse

Create a value of CreateEntityRecognizerResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:entityRecognizerArn:CreateEntityRecognizerResponse', createEntityRecognizerResponse_entityRecognizerArn - The Amazon Resource Name (ARN) that identifies the entity recognizer.

$sel:httpStatus:CreateEntityRecognizerResponse', createEntityRecognizerResponse_httpStatus - The response's http status code.

Response Lenses

createEntityRecognizerResponse_entityRecognizerArn :: Lens' CreateEntityRecognizerResponse (Maybe Text)

The Amazon Resource Name (ARN) that identifies the entity recognizer.

createEntityRecognizerResponse_httpStatus :: Lens' CreateEntityRecognizerResponse Int

The response's http status code.