Amazonka.Comprehend.DescribeEventsDetectionJob

Description

Gets the status and details of an events detection job.

Creating a Request

data DescribeEventsDetectionJob

See: newDescribeEventsDetectionJob smart constructor.

Constructors

DescribeEventsDetectionJob'

Fields

jobId :: Text

The identifier of the events detection job.

newDescribeEventsDetectionJob

Arguments

:: Text
$sel:jobId:DescribeEventsDetectionJob'

-> DescribeEventsDetectionJob

Create a value of DescribeEventsDetectionJob with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:jobId:DescribeEventsDetectionJob', describeEventsDetectionJob_jobId - The identifier of the events detection job.

Request Lenses

describeEventsDetectionJob_jobId :: Lens' DescribeEventsDetectionJob Text

The identifier of the events detection job.

Destructuring the Response

data DescribeEventsDetectionJobResponse

See: newDescribeEventsDetectionJobResponse smart constructor.

Constructors

DescribeEventsDetectionJobResponse'

Fields

eventsDetectionJobProperties :: Maybe EventsDetectionJobProperties

An object that contains the properties associated with an event detection job.

httpStatus :: Int

The response's http status code.

newDescribeEventsDetectionJobResponse

Arguments

:: Int
$sel:httpStatus:DescribeEventsDetectionJobResponse'

-> DescribeEventsDetectionJobResponse

Create a value of DescribeEventsDetectionJobResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:eventsDetectionJobProperties:DescribeEventsDetectionJobResponse', describeEventsDetectionJobResponse_eventsDetectionJobProperties - An object that contains the properties associated with an event detection job.

$sel:httpStatus:DescribeEventsDetectionJobResponse', describeEventsDetectionJobResponse_httpStatus - The response's http status code.

Response Lenses

describeEventsDetectionJobResponse_eventsDetectionJobProperties :: Lens' DescribeEventsDetectionJobResponse (Maybe EventsDetectionJobProperties)

An object that contains the properties associated with an event detection job.

describeEventsDetectionJobResponse_httpStatus :: Lens' DescribeEventsDetectionJobResponse Int

The response's http status code.