Amazonka.Comprehend.ListEventsDetectionJobs

Description

Gets a list of the events detection jobs that you have submitted.

Creating a Request

data ListEventsDetectionJobs

See: newListEventsDetectionJobs smart constructor.

Constructors

ListEventsDetectionJobs'

Fields

nextToken :: Maybe Text

Identifies the next page of results to return.

filter' :: Maybe EventsDetectionJobFilter

Filters the jobs that are returned. You can filter jobs on their name, status, or the date and time that they were submitted. You can only set one filter at a time.

maxResults :: Maybe Natural

The maximum number of results to return in each page.

newListEventsDetectionJobs :: ListEventsDetectionJobs

Create a value of ListEventsDetectionJobs with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:nextToken:ListEventsDetectionJobs', listEventsDetectionJobs_nextToken - Identifies the next page of results to return.

$sel:filter':ListEventsDetectionJobs', listEventsDetectionJobs_filter - Filters the jobs that are returned. You can filter jobs on their name, status, or the date and time that they were submitted. You can only set one filter at a time.

$sel:maxResults:ListEventsDetectionJobs', listEventsDetectionJobs_maxResults - The maximum number of results to return in each page.

Request Lenses

listEventsDetectionJobs_nextToken :: Lens' ListEventsDetectionJobs (Maybe Text)

Identifies the next page of results to return.

listEventsDetectionJobs_filter :: Lens' ListEventsDetectionJobs (Maybe EventsDetectionJobFilter)

Filters the jobs that are returned. You can filter jobs on their name, status, or the date and time that they were submitted. You can only set one filter at a time.

listEventsDetectionJobs_maxResults :: Lens' ListEventsDetectionJobs (Maybe Natural)

The maximum number of results to return in each page.

Destructuring the Response

data ListEventsDetectionJobsResponse

See: newListEventsDetectionJobsResponse smart constructor.

Constructors

ListEventsDetectionJobsResponse'

Fields

eventsDetectionJobPropertiesList :: Maybe [EventsDetectionJobProperties]

A list containing the properties of each job that is returned.

nextToken :: Maybe Text

Identifies the next page of results to return.

httpStatus :: Int

The response's http status code.

newListEventsDetectionJobsResponse

Arguments

:: Int
$sel:httpStatus:ListEventsDetectionJobsResponse'

-> ListEventsDetectionJobsResponse

Create a value of ListEventsDetectionJobsResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:eventsDetectionJobPropertiesList:ListEventsDetectionJobsResponse', listEventsDetectionJobsResponse_eventsDetectionJobPropertiesList - A list containing the properties of each job that is returned.

$sel:nextToken:ListEventsDetectionJobs', listEventsDetectionJobsResponse_nextToken - Identifies the next page of results to return.

$sel:httpStatus:ListEventsDetectionJobsResponse', listEventsDetectionJobsResponse_httpStatus - The response's http status code.

Response Lenses

listEventsDetectionJobsResponse_eventsDetectionJobPropertiesList :: Lens' ListEventsDetectionJobsResponse (Maybe [EventsDetectionJobProperties])

A list containing the properties of each job that is returned.

listEventsDetectionJobsResponse_nextToken :: Lens' ListEventsDetectionJobsResponse (Maybe Text)

Identifies the next page of results to return.

listEventsDetectionJobsResponse_httpStatus :: Lens' ListEventsDetectionJobsResponse Int

The response's http status code.