Amazonka.Comprehend.StartDocumentClassificationJob

Description

Starts an asynchronous document classification job. Use the operation to track the progress of the job.

Creating a Request

data StartDocumentClassificationJob

See: newStartDocumentClassificationJob smart constructor.

Constructors

StartDocumentClassificationJob'

Fields

jobName :: Maybe Text

The identifier of the job.

vpcConfig :: Maybe VpcConfig

Configuration parameters for an optional private Virtual Private Cloud (VPC) containing the resources you are using for your document classification job. For more information, see Amazon VPC.

volumeKmsKeyId :: Maybe Text

ID for the AWS Key Management Service (KMS) key that Amazon Comprehend uses to encrypt data on the storage volume attached to the ML compute instance(s) that process the analysis job. The VolumeKmsKeyId can be either of the following formats:

KMS Key ID: "1234abcd-12ab-34cd-56ef-1234567890ab"

Amazon Resource Name (ARN) of a KMS Key: "arn:aws:kms:us-west-2:111122223333:key/1234abcd-12ab-34cd-56ef-1234567890ab"

clientRequestToken :: Maybe Text

A unique identifier for the request. If you do not set the client request token, Amazon Comprehend generates one.

tags :: Maybe [Tag]

Tags to be associated with the document classification job. A tag is a key-value pair that adds metadata to a resource used by Amazon Comprehend. For example, a tag with "Sales" as the key might be added to a resource to indicate its use by the sales department.

documentClassifierArn :: Text

The Amazon Resource Name (ARN) of the document classifier to use to process the job.

inputDataConfig :: InputDataConfig

Specifies the format and location of the input data for the job.

outputDataConfig :: OutputDataConfig

Specifies where to send the output files.

dataAccessRoleArn :: Text

The Amazon Resource Name (ARN) of the AWS Identity and Access Management (IAM) role that grants Amazon Comprehend read access to your input data.

newStartDocumentClassificationJob

Arguments

:: Text
$sel:documentClassifierArn:StartDocumentClassificationJob'

-> InputDataConfig
$sel:inputDataConfig:StartDocumentClassificationJob'

-> OutputDataConfig
$sel:outputDataConfig:StartDocumentClassificationJob'

-> Text
$sel:dataAccessRoleArn:StartDocumentClassificationJob'

-> StartDocumentClassificationJob

Create a value of StartDocumentClassificationJob with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:jobName:StartDocumentClassificationJob', startDocumentClassificationJob_jobName - The identifier of the job.

$sel:vpcConfig:StartDocumentClassificationJob', startDocumentClassificationJob_vpcConfig - Configuration parameters for an optional private Virtual Private Cloud (VPC) containing the resources you are using for your document classification job. For more information, see Amazon VPC.

$sel:volumeKmsKeyId:StartDocumentClassificationJob', startDocumentClassificationJob_volumeKmsKeyId - ID for the AWS Key Management Service (KMS) key that Amazon Comprehend uses to encrypt data on the storage volume attached to the ML compute instance(s) that process the analysis job. The VolumeKmsKeyId can be either of the following formats:

KMS Key ID: "1234abcd-12ab-34cd-56ef-1234567890ab"

Amazon Resource Name (ARN) of a KMS Key: "arn:aws:kms:us-west-2:111122223333:key/1234abcd-12ab-34cd-56ef-1234567890ab"

$sel:clientRequestToken:StartDocumentClassificationJob', startDocumentClassificationJob_clientRequestToken - A unique identifier for the request. If you do not set the client request token, Amazon Comprehend generates one.

$sel:tags:StartDocumentClassificationJob', startDocumentClassificationJob_tags - Tags to be associated with the document classification job. A tag is a key-value pair that adds metadata to a resource used by Amazon Comprehend. For example, a tag with "Sales" as the key might be added to a resource to indicate its use by the sales department.

$sel:documentClassifierArn:StartDocumentClassificationJob', startDocumentClassificationJob_documentClassifierArn - The Amazon Resource Name (ARN) of the document classifier to use to process the job.

$sel:inputDataConfig:StartDocumentClassificationJob', startDocumentClassificationJob_inputDataConfig - Specifies the format and location of the input data for the job.

$sel:outputDataConfig:StartDocumentClassificationJob', startDocumentClassificationJob_outputDataConfig - Specifies where to send the output files.

$sel:dataAccessRoleArn:StartDocumentClassificationJob', startDocumentClassificationJob_dataAccessRoleArn - The Amazon Resource Name (ARN) of the AWS Identity and Access Management (IAM) role that grants Amazon Comprehend read access to your input data.

Request Lenses

startDocumentClassificationJob_jobName :: Lens' StartDocumentClassificationJob (Maybe Text)

The identifier of the job.

startDocumentClassificationJob_vpcConfig :: Lens' StartDocumentClassificationJob (Maybe VpcConfig)

Configuration parameters for an optional private Virtual Private Cloud (VPC) containing the resources you are using for your document classification job. For more information, see Amazon VPC.

startDocumentClassificationJob_volumeKmsKeyId :: Lens' StartDocumentClassificationJob (Maybe Text)

ID for the AWS Key Management Service (KMS) key that Amazon Comprehend uses to encrypt data on the storage volume attached to the ML compute instance(s) that process the analysis job. The VolumeKmsKeyId can be either of the following formats:

KMS Key ID: "1234abcd-12ab-34cd-56ef-1234567890ab"

Amazon Resource Name (ARN) of a KMS Key: "arn:aws:kms:us-west-2:111122223333:key/1234abcd-12ab-34cd-56ef-1234567890ab"

startDocumentClassificationJob_clientRequestToken :: Lens' StartDocumentClassificationJob (Maybe Text)

A unique identifier for the request. If you do not set the client request token, Amazon Comprehend generates one.

startDocumentClassificationJob_tags :: Lens' StartDocumentClassificationJob (Maybe [Tag])

Tags to be associated with the document classification job. A tag is a key-value pair that adds metadata to a resource used by Amazon Comprehend. For example, a tag with "Sales" as the key might be added to a resource to indicate its use by the sales department.

startDocumentClassificationJob_documentClassifierArn :: Lens' StartDocumentClassificationJob Text

The Amazon Resource Name (ARN) of the document classifier to use to process the job.

startDocumentClassificationJob_inputDataConfig :: Lens' StartDocumentClassificationJob InputDataConfig

Specifies the format and location of the input data for the job.

startDocumentClassificationJob_outputDataConfig :: Lens' StartDocumentClassificationJob OutputDataConfig

Specifies where to send the output files.

startDocumentClassificationJob_dataAccessRoleArn :: Lens' StartDocumentClassificationJob Text

The Amazon Resource Name (ARN) of the AWS Identity and Access Management (IAM) role that grants Amazon Comprehend read access to your input data.

Destructuring the Response

data StartDocumentClassificationJobResponse

See: newStartDocumentClassificationJobResponse smart constructor.

Constructors

StartDocumentClassificationJobResponse'

Fields

jobId :: Maybe Text

The identifier generated for the job. To get the status of the job, use this identifier with the operation.

jobArn :: Maybe Text

The Amazon Resource Name (ARN) of the document classification job. It is a unique, fully qualified identifier for the job. It includes the AWS account, Region, and the job ID. The format of the ARN is as follows:

arn:<partition>:comprehend:<region>:<account-id>:document-classification-job/<job-id>

The following is an example job ARN:

arn:aws:comprehend:us-west-2:111122223333:document-classification-job/1234abcd12ab34cd56ef1234567890ab

jobStatus :: Maybe JobStatus

The status of the job:

SUBMITTED - The job has been received and queued for processing.

IN_PROGRESS - Amazon Comprehend is processing the job.

COMPLETED - The job was successfully completed and the output is available.

FAILED - The job did not complete. For details, use the operation.

STOP_REQUESTED - Amazon Comprehend has received a stop request for the job and is processing the request.

STOPPED - The job was successfully stopped without completing.

httpStatus :: Int

The response's http status code.

newStartDocumentClassificationJobResponse

Arguments

:: Int
$sel:httpStatus:StartDocumentClassificationJobResponse'

-> StartDocumentClassificationJobResponse

Create a value of StartDocumentClassificationJobResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:jobId:StartDocumentClassificationJobResponse', startDocumentClassificationJobResponse_jobId - The identifier generated for the job. To get the status of the job, use this identifier with the operation.

$sel:jobArn:StartDocumentClassificationJobResponse', startDocumentClassificationJobResponse_jobArn - The Amazon Resource Name (ARN) of the document classification job. It is a unique, fully qualified identifier for the job. It includes the AWS account, Region, and the job ID. The format of the ARN is as follows:

arn:<partition>:comprehend:<region>:<account-id>:document-classification-job/<job-id>

The following is an example job ARN:

arn:aws:comprehend:us-west-2:111122223333:document-classification-job/1234abcd12ab34cd56ef1234567890ab

$sel:jobStatus:StartDocumentClassificationJobResponse', startDocumentClassificationJobResponse_jobStatus - The status of the job:

SUBMITTED - The job has been received and queued for processing.

IN_PROGRESS - Amazon Comprehend is processing the job.

COMPLETED - The job was successfully completed and the output is available.

FAILED - The job did not complete. For details, use the operation.

STOP_REQUESTED - Amazon Comprehend has received a stop request for the job and is processing the request.

STOPPED - The job was successfully stopped without completing.

$sel:httpStatus:StartDocumentClassificationJobResponse', startDocumentClassificationJobResponse_httpStatus - The response's http status code.

Response Lenses

startDocumentClassificationJobResponse_jobId :: Lens' StartDocumentClassificationJobResponse (Maybe Text)

The identifier generated for the job. To get the status of the job, use this identifier with the operation.

startDocumentClassificationJobResponse_jobArn :: Lens' StartDocumentClassificationJobResponse (Maybe Text)

The Amazon Resource Name (ARN) of the document classification job. It is a unique, fully qualified identifier for the job. It includes the AWS account, Region, and the job ID. The format of the ARN is as follows:

arn:<partition>:comprehend:<region>:<account-id>:document-classification-job/<job-id>

The following is an example job ARN:

arn:aws:comprehend:us-west-2:111122223333:document-classification-job/1234abcd12ab34cd56ef1234567890ab

startDocumentClassificationJobResponse_jobStatus :: Lens' StartDocumentClassificationJobResponse (Maybe JobStatus)

The status of the job:

SUBMITTED - The job has been received and queued for processing.

IN_PROGRESS - Amazon Comprehend is processing the job.

COMPLETED - The job was successfully completed and the output is available.

FAILED - The job did not complete. For details, use the operation.

STOP_REQUESTED - Amazon Comprehend has received a stop request for the job and is processing the request.

STOPPED - The job was successfully stopped without completing.

startDocumentClassificationJobResponse_httpStatus :: Lens' StartDocumentClassificationJobResponse Int

The response's http status code.