Amazonka.Comprehend.StartPiiEntitiesDetectionJob

Description

Starts an asynchronous PII entity detection job for a collection of documents.

Creating a Request

data StartPiiEntitiesDetectionJob

See: newStartPiiEntitiesDetectionJob smart constructor.

Constructors

StartPiiEntitiesDetectionJob'

Fields

jobName :: Maybe Text

The identifier of the job.

redactionConfig :: Maybe RedactionConfig

Provides configuration parameters for PII entity redaction.

This parameter is required if you set the Mode parameter to ONLY_REDACTION. In that case, you must provide a RedactionConfig definition that includes the PiiEntityTypes parameter.

clientRequestToken :: Maybe Text

A unique identifier for the request. If you don't set the client request token, Amazon Comprehend generates one.

tags :: Maybe [Tag]

Tags to be associated with the PII entities detection job. A tag is a key-value pair that adds metadata to a resource used by Amazon Comprehend. For example, a tag with "Sales" as the key might be added to a resource to indicate its use by the sales department.

inputDataConfig :: InputDataConfig

The input properties for a PII entities detection job.

outputDataConfig :: OutputDataConfig

Provides conﬁguration parameters for the output of PII entity detection jobs.

mode :: PiiEntitiesDetectionMode

Specifies whether the output provides the locations (offsets) of PII entities or a file in which PII entities are redacted.

dataAccessRoleArn :: Text

The Amazon Resource Name (ARN) of the AWS Identity and Access Management (IAM) role that grants Amazon Comprehend read access to your input data.

languageCode :: LanguageCode

The language of the input documents.

newStartPiiEntitiesDetectionJob

Arguments

:: InputDataConfig
$sel:inputDataConfig:StartPiiEntitiesDetectionJob'

-> OutputDataConfig
$sel:outputDataConfig:StartPiiEntitiesDetectionJob'

-> PiiEntitiesDetectionMode
$sel:mode:StartPiiEntitiesDetectionJob'

-> Text
$sel:dataAccessRoleArn:StartPiiEntitiesDetectionJob'

-> LanguageCode
$sel:languageCode:StartPiiEntitiesDetectionJob'

-> StartPiiEntitiesDetectionJob

Create a value of StartPiiEntitiesDetectionJob with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:jobName:StartPiiEntitiesDetectionJob', startPiiEntitiesDetectionJob_jobName - The identifier of the job.

$sel:redactionConfig:StartPiiEntitiesDetectionJob', startPiiEntitiesDetectionJob_redactionConfig - Provides configuration parameters for PII entity redaction.

This parameter is required if you set the Mode parameter to ONLY_REDACTION. In that case, you must provide a RedactionConfig definition that includes the PiiEntityTypes parameter.

$sel:clientRequestToken:StartPiiEntitiesDetectionJob', startPiiEntitiesDetectionJob_clientRequestToken - A unique identifier for the request. If you don't set the client request token, Amazon Comprehend generates one.

$sel:tags:StartPiiEntitiesDetectionJob', startPiiEntitiesDetectionJob_tags - Tags to be associated with the PII entities detection job. A tag is a key-value pair that adds metadata to a resource used by Amazon Comprehend. For example, a tag with "Sales" as the key might be added to a resource to indicate its use by the sales department.

$sel:inputDataConfig:StartPiiEntitiesDetectionJob', startPiiEntitiesDetectionJob_inputDataConfig - The input properties for a PII entities detection job.

$sel:outputDataConfig:StartPiiEntitiesDetectionJob', startPiiEntitiesDetectionJob_outputDataConfig - Provides conﬁguration parameters for the output of PII entity detection jobs.

$sel:mode:StartPiiEntitiesDetectionJob', startPiiEntitiesDetectionJob_mode - Specifies whether the output provides the locations (offsets) of PII entities or a file in which PII entities are redacted.

$sel:dataAccessRoleArn:StartPiiEntitiesDetectionJob', startPiiEntitiesDetectionJob_dataAccessRoleArn - The Amazon Resource Name (ARN) of the AWS Identity and Access Management (IAM) role that grants Amazon Comprehend read access to your input data.

$sel:languageCode:StartPiiEntitiesDetectionJob', startPiiEntitiesDetectionJob_languageCode - The language of the input documents.

Request Lenses

startPiiEntitiesDetectionJob_jobName :: Lens' StartPiiEntitiesDetectionJob (Maybe Text)

The identifier of the job.

startPiiEntitiesDetectionJob_redactionConfig :: Lens' StartPiiEntitiesDetectionJob (Maybe RedactionConfig)

Provides configuration parameters for PII entity redaction.

This parameter is required if you set the Mode parameter to ONLY_REDACTION. In that case, you must provide a RedactionConfig definition that includes the PiiEntityTypes parameter.

startPiiEntitiesDetectionJob_clientRequestToken :: Lens' StartPiiEntitiesDetectionJob (Maybe Text)

A unique identifier for the request. If you don't set the client request token, Amazon Comprehend generates one.

startPiiEntitiesDetectionJob_tags :: Lens' StartPiiEntitiesDetectionJob (Maybe [Tag])

Tags to be associated with the PII entities detection job. A tag is a key-value pair that adds metadata to a resource used by Amazon Comprehend. For example, a tag with "Sales" as the key might be added to a resource to indicate its use by the sales department.

startPiiEntitiesDetectionJob_inputDataConfig :: Lens' StartPiiEntitiesDetectionJob InputDataConfig

The input properties for a PII entities detection job.

startPiiEntitiesDetectionJob_outputDataConfig :: Lens' StartPiiEntitiesDetectionJob OutputDataConfig

Provides conﬁguration parameters for the output of PII entity detection jobs.

startPiiEntitiesDetectionJob_mode :: Lens' StartPiiEntitiesDetectionJob PiiEntitiesDetectionMode

Specifies whether the output provides the locations (offsets) of PII entities or a file in which PII entities are redacted.

startPiiEntitiesDetectionJob_dataAccessRoleArn :: Lens' StartPiiEntitiesDetectionJob Text

The Amazon Resource Name (ARN) of the AWS Identity and Access Management (IAM) role that grants Amazon Comprehend read access to your input data.

startPiiEntitiesDetectionJob_languageCode :: Lens' StartPiiEntitiesDetectionJob LanguageCode

The language of the input documents.

Destructuring the Response

data StartPiiEntitiesDetectionJobResponse

See: newStartPiiEntitiesDetectionJobResponse smart constructor.

Constructors

StartPiiEntitiesDetectionJobResponse'

Fields

jobId :: Maybe Text

The identifier generated for the job.

jobArn :: Maybe Text

The Amazon Resource Name (ARN) of the PII entity detection job. It is a unique, fully qualified identifier for the job. It includes the AWS account, Region, and the job ID. The format of the ARN is as follows:

arn:<partition>:comprehend:<region>:<account-id>:pii-entities-detection-job/<job-id>

The following is an example job ARN:

arn:aws:comprehend:us-west-2:111122223333:pii-entities-detection-job/1234abcd12ab34cd56ef1234567890ab

jobStatus :: Maybe JobStatus

The status of the job.

httpStatus :: Int

The response's http status code.

newStartPiiEntitiesDetectionJobResponse

Arguments

:: Int
$sel:httpStatus:StartPiiEntitiesDetectionJobResponse'

-> StartPiiEntitiesDetectionJobResponse

Create a value of StartPiiEntitiesDetectionJobResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:jobId:StartPiiEntitiesDetectionJobResponse', startPiiEntitiesDetectionJobResponse_jobId - The identifier generated for the job.

$sel:jobArn:StartPiiEntitiesDetectionJobResponse', startPiiEntitiesDetectionJobResponse_jobArn - The Amazon Resource Name (ARN) of the PII entity detection job. It is a unique, fully qualified identifier for the job. It includes the AWS account, Region, and the job ID. The format of the ARN is as follows:

arn:<partition>:comprehend:<region>:<account-id>:pii-entities-detection-job/<job-id>

The following is an example job ARN:

arn:aws:comprehend:us-west-2:111122223333:pii-entities-detection-job/1234abcd12ab34cd56ef1234567890ab

$sel:jobStatus:StartPiiEntitiesDetectionJobResponse', startPiiEntitiesDetectionJobResponse_jobStatus - The status of the job.

$sel:httpStatus:StartPiiEntitiesDetectionJobResponse', startPiiEntitiesDetectionJobResponse_httpStatus - The response's http status code.

Response Lenses

startPiiEntitiesDetectionJobResponse_jobId :: Lens' StartPiiEntitiesDetectionJobResponse (Maybe Text)

The identifier generated for the job.

startPiiEntitiesDetectionJobResponse_jobArn :: Lens' StartPiiEntitiesDetectionJobResponse (Maybe Text)

The Amazon Resource Name (ARN) of the PII entity detection job. It is a unique, fully qualified identifier for the job. It includes the AWS account, Region, and the job ID. The format of the ARN is as follows:

arn:<partition>:comprehend:<region>:<account-id>:pii-entities-detection-job/<job-id>

The following is an example job ARN:

arn:aws:comprehend:us-west-2:111122223333:pii-entities-detection-job/1234abcd12ab34cd56ef1234567890ab

startPiiEntitiesDetectionJobResponse_jobStatus :: Lens' StartPiiEntitiesDetectionJobResponse (Maybe JobStatus)

The status of the job.

startPiiEntitiesDetectionJobResponse_httpStatus :: Lens' StartPiiEntitiesDetectionJobResponse Int

The response's http status code.