Amazonka.Comprehend.StopEntitiesDetectionJob

Description

Stops an entities detection job in progress.

If the job state is IN_PROGRESS the job is marked for termination and put into the STOP_REQUESTED state. If the job completes before it can be stopped, it is put into the COMPLETED state; otherwise the job is stopped and put into the STOPPED state.

If the job is in the COMPLETED or FAILED state when you call the StopDominantLanguageDetectionJob operation, the operation returns a 400 Internal Request Exception.

When a job is stopped, any documents already processed are written to the output location.

Creating a Request

data StopEntitiesDetectionJob

See: newStopEntitiesDetectionJob smart constructor.

Constructors

StopEntitiesDetectionJob'

Fields

jobId :: Text

The identifier of the entities detection job to stop.

newStopEntitiesDetectionJob

Arguments

:: Text
$sel:jobId:StopEntitiesDetectionJob'

-> StopEntitiesDetectionJob

Create a value of StopEntitiesDetectionJob with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:jobId:StopEntitiesDetectionJob', stopEntitiesDetectionJob_jobId - The identifier of the entities detection job to stop.

Request Lenses

stopEntitiesDetectionJob_jobId :: Lens' StopEntitiesDetectionJob Text

The identifier of the entities detection job to stop.

Destructuring the Response

data StopEntitiesDetectionJobResponse

See: newStopEntitiesDetectionJobResponse smart constructor.

Constructors

StopEntitiesDetectionJobResponse'

Fields

jobId :: Maybe Text

The identifier of the entities detection job to stop.

jobStatus :: Maybe JobStatus

Either STOP_REQUESTED if the job is currently running, or STOPPED if the job was previously stopped with the StopEntitiesDetectionJob operation.

httpStatus :: Int

The response's http status code.

newStopEntitiesDetectionJobResponse

Arguments

:: Int
$sel:httpStatus:StopEntitiesDetectionJobResponse'

-> StopEntitiesDetectionJobResponse

Create a value of StopEntitiesDetectionJobResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:jobId:StopEntitiesDetectionJob', stopEntitiesDetectionJobResponse_jobId - The identifier of the entities detection job to stop.

$sel:jobStatus:StopEntitiesDetectionJobResponse', stopEntitiesDetectionJobResponse_jobStatus - Either STOP_REQUESTED if the job is currently running, or STOPPED if the job was previously stopped with the StopEntitiesDetectionJob operation.

$sel:httpStatus:StopEntitiesDetectionJobResponse', stopEntitiesDetectionJobResponse_httpStatus - The response's http status code.

Response Lenses

stopEntitiesDetectionJobResponse_jobId :: Lens' StopEntitiesDetectionJobResponse (Maybe Text)

The identifier of the entities detection job to stop.

stopEntitiesDetectionJobResponse_jobStatus :: Lens' StopEntitiesDetectionJobResponse (Maybe JobStatus)

Either STOP_REQUESTED if the job is currently running, or STOPPED if the job was previously stopped with the StopEntitiesDetectionJob operation.

stopEntitiesDetectionJobResponse_httpStatus :: Lens' StopEntitiesDetectionJobResponse Int

The response's http status code.