Amazonka.Comprehend.Types.BatchDetectDominantLanguageItemResult

Description

Documentation

data BatchDetectDominantLanguageItemResult

The result of calling the operation. The operation returns one object for each document that is successfully processed by the operation.

See: newBatchDetectDominantLanguageItemResult smart constructor.

Constructors

BatchDetectDominantLanguageItemResult'

Fields

languages :: Maybe [DominantLanguage]

One or more DominantLanguage objects describing the dominant languages in the document.

index :: Maybe Int

The zero-based index of the document in the input list.

newBatchDetectDominantLanguageItemResult :: BatchDetectDominantLanguageItemResult

Create a value of BatchDetectDominantLanguageItemResult with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:languages:BatchDetectDominantLanguageItemResult', batchDetectDominantLanguageItemResult_languages - One or more DominantLanguage objects describing the dominant languages in the document.

$sel:index:BatchDetectDominantLanguageItemResult', batchDetectDominantLanguageItemResult_index - The zero-based index of the document in the input list.

batchDetectDominantLanguageItemResult_languages :: Lens' BatchDetectDominantLanguageItemResult (Maybe [DominantLanguage])

One or more DominantLanguage objects describing the dominant languages in the document.

batchDetectDominantLanguageItemResult_index :: Lens' BatchDetectDominantLanguageItemResult (Maybe Int)

The zero-based index of the document in the input list.