Amazonka.Comprehend.Types.EntityRecognizerAnnotations

Description

Documentation

data EntityRecognizerAnnotations

Describes the annotations associated with a entity recognizer.

See: newEntityRecognizerAnnotations smart constructor.

Constructors

EntityRecognizerAnnotations'

Fields

testS3Uri :: Maybe Text

This specifies the Amazon S3 location where the test annotations for an entity recognizer are located. The URI must be in the same AWS Region as the API endpoint that you are calling.

s3Uri :: Text

Specifies the Amazon S3 location where the annotations for an entity recognizer are located. The URI must be in the same region as the API endpoint that you are calling.

newEntityRecognizerAnnotations

Arguments

:: Text
$sel:s3Uri:EntityRecognizerAnnotations'

-> EntityRecognizerAnnotations

Create a value of EntityRecognizerAnnotations with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:testS3Uri:EntityRecognizerAnnotations', entityRecognizerAnnotations_testS3Uri - This specifies the Amazon S3 location where the test annotations for an entity recognizer are located. The URI must be in the same AWS Region as the API endpoint that you are calling.

$sel:s3Uri:EntityRecognizerAnnotations', entityRecognizerAnnotations_s3Uri - Specifies the Amazon S3 location where the annotations for an entity recognizer are located. The URI must be in the same region as the API endpoint that you are calling.

entityRecognizerAnnotations_testS3Uri :: Lens' EntityRecognizerAnnotations (Maybe Text)

This specifies the Amazon S3 location where the test annotations for an entity recognizer are located. The URI must be in the same AWS Region as the API endpoint that you are calling.

entityRecognizerAnnotations_s3Uri :: Lens' EntityRecognizerAnnotations Text

Specifies the Amazon S3 location where the annotations for an entity recognizer are located. The URI must be in the same region as the API endpoint that you are calling.