Amazonka.Comprehend.Types.EntityTypesEvaluationMetrics

Description

Documentation

data EntityTypesEvaluationMetrics

Detailed information about the accuracy of an entity recognizer for a specific entity type.

See: newEntityTypesEvaluationMetrics smart constructor.

Constructors

EntityTypesEvaluationMetrics'

Fields

recall :: Maybe Double

A measure of how complete the recognizer results are for a specific entity type in the test data. High recall means that the recognizer returned most of the relevant results.

precision :: Maybe Double

A measure of the usefulness of the recognizer results for a specific entity type in the test data. High precision means that the recognizer returned substantially more relevant results than irrelevant ones.

f1Score :: Maybe Double

A measure of how accurate the recognizer results are for a specific entity type in the test data. It is derived from the Precision and Recall values. The F1Score is the harmonic average of the two scores. The highest score is 1, and the worst score is 0.

newEntityTypesEvaluationMetrics :: EntityTypesEvaluationMetrics

Create a value of EntityTypesEvaluationMetrics with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:recall:EntityTypesEvaluationMetrics', entityTypesEvaluationMetrics_recall - A measure of how complete the recognizer results are for a specific entity type in the test data. High recall means that the recognizer returned most of the relevant results.

$sel:precision:EntityTypesEvaluationMetrics', entityTypesEvaluationMetrics_precision - A measure of the usefulness of the recognizer results for a specific entity type in the test data. High precision means that the recognizer returned substantially more relevant results than irrelevant ones.

$sel:f1Score:EntityTypesEvaluationMetrics', entityTypesEvaluationMetrics_f1Score - A measure of how accurate the recognizer results are for a specific entity type in the test data. It is derived from the Precision and Recall values. The F1Score is the harmonic average of the two scores. The highest score is 1, and the worst score is 0.

entityTypesEvaluationMetrics_recall :: Lens' EntityTypesEvaluationMetrics (Maybe Double)

A measure of how complete the recognizer results are for a specific entity type in the test data. High recall means that the recognizer returned most of the relevant results.

entityTypesEvaluationMetrics_precision :: Lens' EntityTypesEvaluationMetrics (Maybe Double)

A measure of the usefulness of the recognizer results for a specific entity type in the test data. High precision means that the recognizer returned substantially more relevant results than irrelevant ones.

entityTypesEvaluationMetrics_f1Score :: Lens' EntityTypesEvaluationMetrics (Maybe Double)

A measure of how accurate the recognizer results are for a specific entity type in the test data. It is derived from the Precision and Recall values. The F1Score is the harmonic average of the two scores. The highest score is 1, and the worst score is 0.