Amazonka.Comprehend.Types.RedactionConfig

Description

Documentation

data RedactionConfig

Provides configuration parameters for PII entity redaction.

See: newRedactionConfig smart constructor.

Constructors

RedactionConfig'

Fields

maskCharacter :: Maybe Text

A character that replaces each character in the redacted PII entity.

maskMode :: Maybe PiiEntitiesDetectionMaskMode

Specifies whether the PII entity is redacted with the mask character or the entity type.

piiEntityTypes :: Maybe [PiiEntityType]

An array of the types of PII entities that Amazon Comprehend detects in the input text for your request.

newRedactionConfig :: RedactionConfig

Create a value of RedactionConfig with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:maskCharacter:RedactionConfig', redactionConfig_maskCharacter - A character that replaces each character in the redacted PII entity.

$sel:maskMode:RedactionConfig', redactionConfig_maskMode - Specifies whether the PII entity is redacted with the mask character or the entity type.

$sel:piiEntityTypes:RedactionConfig', redactionConfig_piiEntityTypes - An array of the types of PII entities that Amazon Comprehend detects in the input text for your request.

redactionConfig_maskCharacter :: Lens' RedactionConfig (Maybe Text)

A character that replaces each character in the redacted PII entity.

redactionConfig_maskMode :: Lens' RedactionConfig (Maybe PiiEntitiesDetectionMaskMode)

Specifies whether the PII entity is redacted with the mask character or the entity type.

redactionConfig_piiEntityTypes :: Lens' RedactionConfig (Maybe [PiiEntityType])

An array of the types of PII entities that Amazon Comprehend detects in the input text for your request.