Amazonka.Comprehend.Types.SyntaxToken

Description

Documentation

data SyntaxToken

Represents a work in the input text that was recognized and assigned a part of speech. There is one syntax token record for each word in the source text.

See: newSyntaxToken smart constructor.

Constructors

SyntaxToken'

Fields

beginOffset :: Maybe Int

The zero-based offset from the beginning of the source text to the first character in the word.

text :: Maybe Text

The word that was recognized in the source text.

tokenId :: Maybe Int

A unique identifier for a token.

endOffset :: Maybe Int

The zero-based offset from the beginning of the source text to the last character in the word.

partOfSpeech :: Maybe PartOfSpeechTag

Provides the part of speech label and the confidence level that Amazon Comprehend has that the part of speech was correctly identified. For more information, see how-syntax.

newSyntaxToken :: SyntaxToken

Create a value of SyntaxToken with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:beginOffset:SyntaxToken', syntaxToken_beginOffset - The zero-based offset from the beginning of the source text to the first character in the word.

$sel:text:SyntaxToken', syntaxToken_text - The word that was recognized in the source text.

$sel:tokenId:SyntaxToken', syntaxToken_tokenId - A unique identifier for a token.

$sel:endOffset:SyntaxToken', syntaxToken_endOffset - The zero-based offset from the beginning of the source text to the last character in the word.

$sel:partOfSpeech:SyntaxToken', syntaxToken_partOfSpeech - Provides the part of speech label and the confidence level that Amazon Comprehend has that the part of speech was correctly identified. For more information, see how-syntax.

syntaxToken_beginOffset :: Lens' SyntaxToken (Maybe Int)

The zero-based offset from the beginning of the source text to the first character in the word.

syntaxToken_text :: Lens' SyntaxToken (Maybe Text)

The word that was recognized in the source text.

syntaxToken_tokenId :: Lens' SyntaxToken (Maybe Int)

A unique identifier for a token.

syntaxToken_endOffset :: Lens' SyntaxToken (Maybe Int)

The zero-based offset from the beginning of the source text to the last character in the word.

syntaxToken_partOfSpeech :: Lens' SyntaxToken (Maybe PartOfSpeechTag)

Provides the part of speech label and the confidence level that Amazon Comprehend has that the part of speech was correctly identified. For more information, see how-syntax.