Amazonka.ComprehendMedical.StartICD10CMInferenceJob

Description

Starts an asynchronous job to detect medical conditions and link them to the ICD-10-CM ontology. Use the DescribeICD10CMInferenceJob operation to track the status of a job.

Creating a Request

data StartICD10CMInferenceJob

See: newStartICD10CMInferenceJob smart constructor.

Constructors

StartICD10CMInferenceJob'

Fields

kmsKey :: Maybe Text

An AWS Key Management Service key to encrypt your output files. If you do not specify a key, the files are written in plain text.

jobName :: Maybe Text

The identifier of the job.

clientRequestToken :: Maybe Text

A unique identifier for the request. If you don't set the client request token, Amazon Comprehend Medical generates one.

inputDataConfig :: InputDataConfig

Specifies the format and location of the input data for the job.

outputDataConfig :: OutputDataConfig

Specifies where to send the output files.

dataAccessRoleArn :: Text

The Amazon Resource Name (ARN) of the AWS Identity and Access Management (IAM) role that grants Amazon Comprehend Medical read access to your input data. For more information, see Role-Based Permissions Required for Asynchronous Operations.

languageCode :: LanguageCode

The language of the input documents. All documents must be in the same language.

newStartICD10CMInferenceJob

Arguments

:: InputDataConfig
$sel:inputDataConfig:StartICD10CMInferenceJob'

-> OutputDataConfig
$sel:outputDataConfig:StartICD10CMInferenceJob'

-> Text
$sel:dataAccessRoleArn:StartICD10CMInferenceJob'

-> LanguageCode
$sel:languageCode:StartICD10CMInferenceJob'

-> StartICD10CMInferenceJob

Create a value of StartICD10CMInferenceJob with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:kmsKey:StartICD10CMInferenceJob', startICD10CMInferenceJob_kmsKey - An AWS Key Management Service key to encrypt your output files. If you do not specify a key, the files are written in plain text.

$sel:jobName:StartICD10CMInferenceJob', startICD10CMInferenceJob_jobName - The identifier of the job.

$sel:clientRequestToken:StartICD10CMInferenceJob', startICD10CMInferenceJob_clientRequestToken - A unique identifier for the request. If you don't set the client request token, Amazon Comprehend Medical generates one.

$sel:inputDataConfig:StartICD10CMInferenceJob', startICD10CMInferenceJob_inputDataConfig - Specifies the format and location of the input data for the job.

$sel:outputDataConfig:StartICD10CMInferenceJob', startICD10CMInferenceJob_outputDataConfig - Specifies where to send the output files.

$sel:dataAccessRoleArn:StartICD10CMInferenceJob', startICD10CMInferenceJob_dataAccessRoleArn - The Amazon Resource Name (ARN) of the AWS Identity and Access Management (IAM) role that grants Amazon Comprehend Medical read access to your input data. For more information, see Role-Based Permissions Required for Asynchronous Operations.

$sel:languageCode:StartICD10CMInferenceJob', startICD10CMInferenceJob_languageCode - The language of the input documents. All documents must be in the same language.

Request Lenses

startICD10CMInferenceJob_kmsKey :: Lens' StartICD10CMInferenceJob (Maybe Text)

An AWS Key Management Service key to encrypt your output files. If you do not specify a key, the files are written in plain text.

startICD10CMInferenceJob_jobName :: Lens' StartICD10CMInferenceJob (Maybe Text)

The identifier of the job.

startICD10CMInferenceJob_clientRequestToken :: Lens' StartICD10CMInferenceJob (Maybe Text)

A unique identifier for the request. If you don't set the client request token, Amazon Comprehend Medical generates one.

startICD10CMInferenceJob_inputDataConfig :: Lens' StartICD10CMInferenceJob InputDataConfig

Specifies the format and location of the input data for the job.

startICD10CMInferenceJob_outputDataConfig :: Lens' StartICD10CMInferenceJob OutputDataConfig

Specifies where to send the output files.

startICD10CMInferenceJob_dataAccessRoleArn :: Lens' StartICD10CMInferenceJob Text

The Amazon Resource Name (ARN) of the AWS Identity and Access Management (IAM) role that grants Amazon Comprehend Medical read access to your input data. For more information, see Role-Based Permissions Required for Asynchronous Operations.

startICD10CMInferenceJob_languageCode :: Lens' StartICD10CMInferenceJob LanguageCode

The language of the input documents. All documents must be in the same language.

Destructuring the Response

data StartICD10CMInferenceJobResponse

See: newStartICD10CMInferenceJobResponse smart constructor.

Constructors

StartICD10CMInferenceJobResponse'

Fields

jobId :: Maybe Text

The identifier generated for the job. To get the status of a job, use this identifier with the StartICD10CMInferenceJob operation.

httpStatus :: Int

The response's http status code.

newStartICD10CMInferenceJobResponse

Arguments

:: Int
$sel:httpStatus:StartICD10CMInferenceJobResponse'

-> StartICD10CMInferenceJobResponse

Create a value of StartICD10CMInferenceJobResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:jobId:StartICD10CMInferenceJobResponse', startICD10CMInferenceJobResponse_jobId - The identifier generated for the job. To get the status of a job, use this identifier with the StartICD10CMInferenceJob operation.

$sel:httpStatus:StartICD10CMInferenceJobResponse', startICD10CMInferenceJobResponse_httpStatus - The response's http status code.

Response Lenses

startICD10CMInferenceJobResponse_jobId :: Lens' StartICD10CMInferenceJobResponse (Maybe Text)

The identifier generated for the job. To get the status of a job, use this identifier with the StartICD10CMInferenceJob operation.

startICD10CMInferenceJobResponse_httpStatus :: Lens' StartICD10CMInferenceJobResponse Int

The response's http status code.