Amazonka.ComprehendMedical.Types.ComprehendMedicalAsyncJobProperties

Description

Documentation

data ComprehendMedicalAsyncJobProperties

Provides information about a detection job.

See: newComprehendMedicalAsyncJobProperties smart constructor.

Constructors

ComprehendMedicalAsyncJobProperties'

Fields

languageCode :: Maybe LanguageCode

The language code of the input documents.

kmsKey :: Maybe Text

The AWS Key Management Service key, if any, used to encrypt the output files.

jobId :: Maybe Text

The identifier assigned to the detection job.

jobName :: Maybe Text

The name that you assigned to the detection job.

manifestFilePath :: Maybe Text

The path to the file that describes the results of a batch job.

inputDataConfig :: Maybe InputDataConfig

The input data configuration that you supplied when you created the detection job.

endTime :: Maybe POSIX

The time that the detection job completed.

outputDataConfig :: Maybe OutputDataConfig

The output data configuration that you supplied when you created the detection job.

modelVersion :: Maybe Text

The version of the model used to analyze the documents. The version number looks like X.X.X. You can use this information to track the model used for a particular batch of documents.

dataAccessRoleArn :: Maybe Text

The Amazon Resource Name (ARN) that gives Amazon Comprehend Medical read access to your input data.

jobStatus :: Maybe JobStatus

The current status of the detection job. If the status is FAILED, the Message field shows the reason for the failure.

message :: Maybe Text

A description of the status of a job.

expirationTime :: Maybe POSIX

The date and time that job metadata is deleted from the server. Output files in your S3 bucket will not be deleted. After the metadata is deleted, the job will no longer appear in the results of the ListEntitiesDetectionV2Job or the ListPHIDetectionJobs operation.

submitTime :: Maybe POSIX

The time that the detection job was submitted for processing.

newComprehendMedicalAsyncJobProperties :: ComprehendMedicalAsyncJobProperties

Create a value of ComprehendMedicalAsyncJobProperties with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:languageCode:ComprehendMedicalAsyncJobProperties', comprehendMedicalAsyncJobProperties_languageCode - The language code of the input documents.

$sel:kmsKey:ComprehendMedicalAsyncJobProperties', comprehendMedicalAsyncJobProperties_kmsKey - The AWS Key Management Service key, if any, used to encrypt the output files.

$sel:jobId:ComprehendMedicalAsyncJobProperties', comprehendMedicalAsyncJobProperties_jobId - The identifier assigned to the detection job.

$sel:jobName:ComprehendMedicalAsyncJobProperties', comprehendMedicalAsyncJobProperties_jobName - The name that you assigned to the detection job.

$sel:manifestFilePath:ComprehendMedicalAsyncJobProperties', comprehendMedicalAsyncJobProperties_manifestFilePath - The path to the file that describes the results of a batch job.

$sel:inputDataConfig:ComprehendMedicalAsyncJobProperties', comprehendMedicalAsyncJobProperties_inputDataConfig - The input data configuration that you supplied when you created the detection job.

$sel:endTime:ComprehendMedicalAsyncJobProperties', comprehendMedicalAsyncJobProperties_endTime - The time that the detection job completed.

$sel:outputDataConfig:ComprehendMedicalAsyncJobProperties', comprehendMedicalAsyncJobProperties_outputDataConfig - The output data configuration that you supplied when you created the detection job.

$sel:modelVersion:ComprehendMedicalAsyncJobProperties', comprehendMedicalAsyncJobProperties_modelVersion - The version of the model used to analyze the documents. The version number looks like X.X.X. You can use this information to track the model used for a particular batch of documents.

$sel:dataAccessRoleArn:ComprehendMedicalAsyncJobProperties', comprehendMedicalAsyncJobProperties_dataAccessRoleArn - The Amazon Resource Name (ARN) that gives Amazon Comprehend Medical read access to your input data.

$sel:jobStatus:ComprehendMedicalAsyncJobProperties', comprehendMedicalAsyncJobProperties_jobStatus - The current status of the detection job. If the status is FAILED, the Message field shows the reason for the failure.

$sel:message:ComprehendMedicalAsyncJobProperties', comprehendMedicalAsyncJobProperties_message - A description of the status of a job.

$sel:expirationTime:ComprehendMedicalAsyncJobProperties', comprehendMedicalAsyncJobProperties_expirationTime - The date and time that job metadata is deleted from the server. Output files in your S3 bucket will not be deleted. After the metadata is deleted, the job will no longer appear in the results of the ListEntitiesDetectionV2Job or the ListPHIDetectionJobs operation.

$sel:submitTime:ComprehendMedicalAsyncJobProperties', comprehendMedicalAsyncJobProperties_submitTime - The time that the detection job was submitted for processing.

comprehendMedicalAsyncJobProperties_languageCode :: Lens' ComprehendMedicalAsyncJobProperties (Maybe LanguageCode)

The language code of the input documents.

comprehendMedicalAsyncJobProperties_kmsKey :: Lens' ComprehendMedicalAsyncJobProperties (Maybe Text)

The AWS Key Management Service key, if any, used to encrypt the output files.

comprehendMedicalAsyncJobProperties_jobId :: Lens' ComprehendMedicalAsyncJobProperties (Maybe Text)

The identifier assigned to the detection job.

comprehendMedicalAsyncJobProperties_jobName :: Lens' ComprehendMedicalAsyncJobProperties (Maybe Text)

The name that you assigned to the detection job.

comprehendMedicalAsyncJobProperties_manifestFilePath :: Lens' ComprehendMedicalAsyncJobProperties (Maybe Text)

The path to the file that describes the results of a batch job.

comprehendMedicalAsyncJobProperties_inputDataConfig :: Lens' ComprehendMedicalAsyncJobProperties (Maybe InputDataConfig)

The input data configuration that you supplied when you created the detection job.

comprehendMedicalAsyncJobProperties_endTime :: Lens' ComprehendMedicalAsyncJobProperties (Maybe UTCTime)

The time that the detection job completed.

comprehendMedicalAsyncJobProperties_outputDataConfig :: Lens' ComprehendMedicalAsyncJobProperties (Maybe OutputDataConfig)

The output data configuration that you supplied when you created the detection job.

comprehendMedicalAsyncJobProperties_modelVersion :: Lens' ComprehendMedicalAsyncJobProperties (Maybe Text)

The version of the model used to analyze the documents. The version number looks like X.X.X. You can use this information to track the model used for a particular batch of documents.

comprehendMedicalAsyncJobProperties_dataAccessRoleArn :: Lens' ComprehendMedicalAsyncJobProperties (Maybe Text)

The Amazon Resource Name (ARN) that gives Amazon Comprehend Medical read access to your input data.

comprehendMedicalAsyncJobProperties_jobStatus :: Lens' ComprehendMedicalAsyncJobProperties (Maybe JobStatus)

The current status of the detection job. If the status is FAILED, the Message field shows the reason for the failure.

comprehendMedicalAsyncJobProperties_message :: Lens' ComprehendMedicalAsyncJobProperties (Maybe Text)

A description of the status of a job.

comprehendMedicalAsyncJobProperties_expirationTime :: Lens' ComprehendMedicalAsyncJobProperties (Maybe UTCTime)

The date and time that job metadata is deleted from the server. Output files in your S3 bucket will not be deleted. After the metadata is deleted, the job will no longer appear in the results of the ListEntitiesDetectionV2Job or the ListPHIDetectionJobs operation.

comprehendMedicalAsyncJobProperties_submitTime :: Lens' ComprehendMedicalAsyncJobProperties (Maybe UTCTime)

The time that the detection job was submitted for processing.