Amazonka.ComprehendMedical.Types.ICD10CMEntity

Description

Documentation

data ICD10CMEntity

The collection of medical entities extracted from the input text and their associated information. For each entity, the response provides the entity text, the entity category, where the entity text begins and ends, and the level of confidence that Amazon Comprehend Medical has in the detection and analysis. Attributes and traits of the entity are also returned.

See: newICD10CMEntity smart constructor.

Constructors

ICD10CMEntity'

Fields

iCD10CMConcepts :: Maybe [ICD10CMConcept]

The ICD-10-CM concepts that the entity could refer to, along with a score indicating the likelihood of the match.

beginOffset :: Maybe Int

The 0-based character offset in the input text that shows where the entity begins. The offset returns the UTF-8 code point in the string.

text :: Maybe Text

The segment of input text that is matched to the detected entity.

category :: Maybe ICD10CMEntityCategory

The category of the entity. InferICD10CM detects entities in the MEDICAL_CONDITION category.

score :: Maybe Double

The level of confidence that Amazon Comprehend Medical has in the accuracy of the detection.

traits :: Maybe [ICD10CMTrait]

Provides Contextual information for the entity. The traits recognized by InferICD10CM are DIAGNOSIS, SIGN, SYMPTOM, and NEGATION.

attributes :: Maybe [ICD10CMAttribute]

The detected attributes that relate to the entity. An extracted segment of the text that is an attribute of an entity, or otherwise related to an entity, such as the nature of a medical condition.

endOffset :: Maybe Int

The 0-based character offset in the input text that shows where the entity ends. The offset returns the UTF-8 code point in the string.

id :: Maybe Int

The numeric identifier for the entity. This is a monotonically increasing id unique within this response rather than a global unique identifier.

type' :: Maybe ICD10CMEntityType

Describes the specific type of entity with category of entities. InferICD10CM detects entities of the type DX_NAME and TIME_EXPRESSION.

newICD10CMEntity :: ICD10CMEntity

Create a value of ICD10CMEntity with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:iCD10CMConcepts:ICD10CMEntity', iCD10CMEntity_iCD10CMConcepts - The ICD-10-CM concepts that the entity could refer to, along with a score indicating the likelihood of the match.

$sel:beginOffset:ICD10CMEntity', iCD10CMEntity_beginOffset - The 0-based character offset in the input text that shows where the entity begins. The offset returns the UTF-8 code point in the string.

$sel:text:ICD10CMEntity', iCD10CMEntity_text - The segment of input text that is matched to the detected entity.

$sel:category:ICD10CMEntity', iCD10CMEntity_category - The category of the entity. InferICD10CM detects entities in the MEDICAL_CONDITION category.

$sel:score:ICD10CMEntity', iCD10CMEntity_score - The level of confidence that Amazon Comprehend Medical has in the accuracy of the detection.

$sel:traits:ICD10CMEntity', iCD10CMEntity_traits - Provides Contextual information for the entity. The traits recognized by InferICD10CM are DIAGNOSIS, SIGN, SYMPTOM, and NEGATION.

$sel:attributes:ICD10CMEntity', iCD10CMEntity_attributes - The detected attributes that relate to the entity. An extracted segment of the text that is an attribute of an entity, or otherwise related to an entity, such as the nature of a medical condition.

$sel:endOffset:ICD10CMEntity', iCD10CMEntity_endOffset - The 0-based character offset in the input text that shows where the entity ends. The offset returns the UTF-8 code point in the string.

$sel:id:ICD10CMEntity', iCD10CMEntity_id - The numeric identifier for the entity. This is a monotonically increasing id unique within this response rather than a global unique identifier.

$sel:type':ICD10CMEntity', iCD10CMEntity_type - Describes the specific type of entity with category of entities. InferICD10CM detects entities of the type DX_NAME and TIME_EXPRESSION.

iCD10CMEntity_iCD10CMConcepts :: Lens' ICD10CMEntity (Maybe [ICD10CMConcept])

The ICD-10-CM concepts that the entity could refer to, along with a score indicating the likelihood of the match.

iCD10CMEntity_beginOffset :: Lens' ICD10CMEntity (Maybe Int)

The 0-based character offset in the input text that shows where the entity begins. The offset returns the UTF-8 code point in the string.

iCD10CMEntity_text :: Lens' ICD10CMEntity (Maybe Text)

The segment of input text that is matched to the detected entity.

iCD10CMEntity_category :: Lens' ICD10CMEntity (Maybe ICD10CMEntityCategory)

The category of the entity. InferICD10CM detects entities in the MEDICAL_CONDITION category.

iCD10CMEntity_score :: Lens' ICD10CMEntity (Maybe Double)

The level of confidence that Amazon Comprehend Medical has in the accuracy of the detection.

iCD10CMEntity_traits :: Lens' ICD10CMEntity (Maybe [ICD10CMTrait])

Provides Contextual information for the entity. The traits recognized by InferICD10CM are DIAGNOSIS, SIGN, SYMPTOM, and NEGATION.

iCD10CMEntity_attributes :: Lens' ICD10CMEntity (Maybe [ICD10CMAttribute])

The detected attributes that relate to the entity. An extracted segment of the text that is an attribute of an entity, or otherwise related to an entity, such as the nature of a medical condition.

iCD10CMEntity_endOffset :: Lens' ICD10CMEntity (Maybe Int)

The 0-based character offset in the input text that shows where the entity ends. The offset returns the UTF-8 code point in the string.

iCD10CMEntity_id :: Lens' ICD10CMEntity (Maybe Int)

The numeric identifier for the entity. This is a monotonically increasing id unique within this response rather than a global unique identifier.

iCD10CMEntity_type :: Lens' ICD10CMEntity (Maybe ICD10CMEntityType)

Describes the specific type of entity with category of entities. InferICD10CM detects entities of the type DX_NAME and TIME_EXPRESSION.