Amazonka.ComprehendMedical.Types.ICD10CMTrait

Description

Documentation

data ICD10CMTrait

Contextual information for the entity. The traits recognized by InferICD10CM are DIAGNOSIS, SIGN, SYMPTOM, and NEGATION.

See: newICD10CMTrait smart constructor.

Constructors

ICD10CMTrait'

Fields

score :: Maybe Double

The level of confidence that Amazon Comprehend Medical has that the segment of text is correctly recognized as a trait.

name :: Maybe ICD10CMTraitName

Provides a name or contextual description about the trait.

newICD10CMTrait :: ICD10CMTrait

Create a value of ICD10CMTrait with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:score:ICD10CMTrait', iCD10CMTrait_score - The level of confidence that Amazon Comprehend Medical has that the segment of text is correctly recognized as a trait.

$sel:name:ICD10CMTrait', iCD10CMTrait_name - Provides a name or contextual description about the trait.

iCD10CMTrait_score :: Lens' ICD10CMTrait (Maybe Double)

The level of confidence that Amazon Comprehend Medical has that the segment of text is correctly recognized as a trait.

iCD10CMTrait_name :: Lens' ICD10CMTrait (Maybe ICD10CMTraitName)

Provides a name or contextual description about the trait.