Amazonka.ComprehendMedical.Types.RxNormConcept

Description

Documentation

data RxNormConcept

The RxNorm concept that the entity could refer to, along with a score indicating the likelihood of the match.

See: newRxNormConcept smart constructor.

Constructors

RxNormConcept'

Fields

score :: Maybe Double

The level of confidence that Amazon Comprehend Medical has that the entity is accurately linked to the reported RxNorm concept.

code :: Maybe Text

RxNorm concept ID, also known as the RxCUI.

description :: Maybe Text

The description of the RxNorm concept.

newRxNormConcept :: RxNormConcept

Create a value of RxNormConcept with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:score:RxNormConcept', rxNormConcept_score - The level of confidence that Amazon Comprehend Medical has that the entity is accurately linked to the reported RxNorm concept.

$sel:code:RxNormConcept', rxNormConcept_code - RxNorm concept ID, also known as the RxCUI.

$sel:description:RxNormConcept', rxNormConcept_description - The description of the RxNorm concept.

rxNormConcept_score :: Lens' RxNormConcept (Maybe Double)

The level of confidence that Amazon Comprehend Medical has that the entity is accurately linked to the reported RxNorm concept.

rxNormConcept_code :: Lens' RxNormConcept (Maybe Text)

RxNorm concept ID, also known as the RxCUI.

rxNormConcept_description :: Lens' RxNormConcept (Maybe Text)

The description of the RxNorm concept.