Amazonka.ComputeOptimizer.ExportAutoScalingGroupRecommendations

Description

Exports optimization recommendations for Auto Scaling groups.

Recommendations are exported in a comma-separated values (.csv) file, and its metadata in a JavaScript Object Notation (JSON) (.json) file, to an existing Amazon Simple Storage Service (Amazon S3) bucket that you specify. For more information, see Exporting Recommendations in the Compute Optimizer User Guide.

You can have only one Auto Scaling group export job in progress per Amazon Web Services Region.

Creating a Request

data ExportAutoScalingGroupRecommendations

See: newExportAutoScalingGroupRecommendations smart constructor.

Constructors

ExportAutoScalingGroupRecommendations'

Fields

accountIds :: Maybe [Text]

The IDs of the Amazon Web Services accounts for which to export Auto Scaling group recommendations.

If your account is the management account of an organization, use this parameter to specify the member account for which you want to export recommendations.

This parameter cannot be specified together with the include member accounts parameter. The parameters are mutually exclusive.

Recommendations for member accounts are not included in the export if this parameter, or the include member accounts parameter, is omitted.

You can specify multiple account IDs per request.

fileFormat :: Maybe FileFormat

The format of the export file.

The only export file format currently supported is Csv.

filters :: Maybe [Filter]

An array of objects to specify a filter that exports a more specific set of Auto Scaling group recommendations.

fieldsToExport :: Maybe [ExportableAutoScalingGroupField]

The recommendations data to include in the export file. For more information about the fields that can be exported, see Exported files in the Compute Optimizer User Guide.

includeMemberAccounts :: Maybe Bool

Indicates whether to include recommendations for resources in all member accounts of the organization if your account is the management account of an organization.

The member accounts must also be opted in to Compute Optimizer, and trusted access for Compute Optimizer must be enabled in the organization account. For more information, see Compute Optimizer and Amazon Web Services Organizations trusted access in the Compute Optimizer User Guide.

Recommendations for member accounts of the organization are not included in the export file if this parameter is omitted.

This parameter cannot be specified together with the account IDs parameter. The parameters are mutually exclusive.

Recommendations for member accounts are not included in the export if this parameter, or the account IDs parameter, is omitted.

recommendationPreferences :: Maybe RecommendationPreferences

An object to specify the preferences for the Auto Scaling group recommendations to export.

s3DestinationConfig :: S3DestinationConfig

An object to specify the destination Amazon Simple Storage Service (Amazon S3) bucket name and key prefix for the export job.

You must create the destination Amazon S3 bucket for your recommendations export before you create the export job. Compute Optimizer does not create the S3 bucket for you. After you create the S3 bucket, ensure that it has the required permissions policy to allow Compute Optimizer to write the export file to it. If you plan to specify an object prefix when you create the export job, you must include the object prefix in the policy that you add to the S3 bucket. For more information, see Amazon S3 Bucket Policy for Compute Optimizer in the Compute Optimizer User Guide.

newExportAutoScalingGroupRecommendations

Arguments

:: S3DestinationConfig
$sel:s3DestinationConfig:ExportAutoScalingGroupRecommendations'

-> ExportAutoScalingGroupRecommendations

Create a value of ExportAutoScalingGroupRecommendations with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:accountIds:ExportAutoScalingGroupRecommendations', exportAutoScalingGroupRecommendations_accountIds - The IDs of the Amazon Web Services accounts for which to export Auto Scaling group recommendations.

If your account is the management account of an organization, use this parameter to specify the member account for which you want to export recommendations.

This parameter cannot be specified together with the include member accounts parameter. The parameters are mutually exclusive.

Recommendations for member accounts are not included in the export if this parameter, or the include member accounts parameter, is omitted.

You can specify multiple account IDs per request.

$sel:fileFormat:ExportAutoScalingGroupRecommendations', exportAutoScalingGroupRecommendations_fileFormat - The format of the export file.

The only export file format currently supported is Csv.

$sel:filters:ExportAutoScalingGroupRecommendations', exportAutoScalingGroupRecommendations_filters - An array of objects to specify a filter that exports a more specific set of Auto Scaling group recommendations.

$sel:fieldsToExport:ExportAutoScalingGroupRecommendations', exportAutoScalingGroupRecommendations_fieldsToExport - The recommendations data to include in the export file. For more information about the fields that can be exported, see Exported files in the Compute Optimizer User Guide.

$sel:includeMemberAccounts:ExportAutoScalingGroupRecommendations', exportAutoScalingGroupRecommendations_includeMemberAccounts - Indicates whether to include recommendations for resources in all member accounts of the organization if your account is the management account of an organization.

The member accounts must also be opted in to Compute Optimizer, and trusted access for Compute Optimizer must be enabled in the organization account. For more information, see Compute Optimizer and Amazon Web Services Organizations trusted access in the Compute Optimizer User Guide.

Recommendations for member accounts of the organization are not included in the export file if this parameter is omitted.

This parameter cannot be specified together with the account IDs parameter. The parameters are mutually exclusive.

Recommendations for member accounts are not included in the export if this parameter, or the account IDs parameter, is omitted.

$sel:recommendationPreferences:ExportAutoScalingGroupRecommendations', exportAutoScalingGroupRecommendations_recommendationPreferences - An object to specify the preferences for the Auto Scaling group recommendations to export.

$sel:s3DestinationConfig:ExportAutoScalingGroupRecommendations', exportAutoScalingGroupRecommendations_s3DestinationConfig - An object to specify the destination Amazon Simple Storage Service (Amazon S3) bucket name and key prefix for the export job.

You must create the destination Amazon S3 bucket for your recommendations export before you create the export job. Compute Optimizer does not create the S3 bucket for you. After you create the S3 bucket, ensure that it has the required permissions policy to allow Compute Optimizer to write the export file to it. If you plan to specify an object prefix when you create the export job, you must include the object prefix in the policy that you add to the S3 bucket. For more information, see Amazon S3 Bucket Policy for Compute Optimizer in the Compute Optimizer User Guide.

Request Lenses

exportAutoScalingGroupRecommendations_accountIds :: Lens' ExportAutoScalingGroupRecommendations (Maybe [Text])

The IDs of the Amazon Web Services accounts for which to export Auto Scaling group recommendations.

If your account is the management account of an organization, use this parameter to specify the member account for which you want to export recommendations.

This parameter cannot be specified together with the include member accounts parameter. The parameters are mutually exclusive.

Recommendations for member accounts are not included in the export if this parameter, or the include member accounts parameter, is omitted.

You can specify multiple account IDs per request.

exportAutoScalingGroupRecommendations_fileFormat :: Lens' ExportAutoScalingGroupRecommendations (Maybe FileFormat)

The format of the export file.

The only export file format currently supported is Csv.

exportAutoScalingGroupRecommendations_filters :: Lens' ExportAutoScalingGroupRecommendations (Maybe [Filter])

An array of objects to specify a filter that exports a more specific set of Auto Scaling group recommendations.

exportAutoScalingGroupRecommendations_fieldsToExport :: Lens' ExportAutoScalingGroupRecommendations (Maybe [ExportableAutoScalingGroupField])

The recommendations data to include in the export file. For more information about the fields that can be exported, see Exported files in the Compute Optimizer User Guide.

exportAutoScalingGroupRecommendations_includeMemberAccounts :: Lens' ExportAutoScalingGroupRecommendations (Maybe Bool)

Indicates whether to include recommendations for resources in all member accounts of the organization if your account is the management account of an organization.

The member accounts must also be opted in to Compute Optimizer, and trusted access for Compute Optimizer must be enabled in the organization account. For more information, see Compute Optimizer and Amazon Web Services Organizations trusted access in the Compute Optimizer User Guide.

Recommendations for member accounts of the organization are not included in the export file if this parameter is omitted.

This parameter cannot be specified together with the account IDs parameter. The parameters are mutually exclusive.

Recommendations for member accounts are not included in the export if this parameter, or the account IDs parameter, is omitted.

exportAutoScalingGroupRecommendations_recommendationPreferences :: Lens' ExportAutoScalingGroupRecommendations (Maybe RecommendationPreferences)

An object to specify the preferences for the Auto Scaling group recommendations to export.

exportAutoScalingGroupRecommendations_s3DestinationConfig :: Lens' ExportAutoScalingGroupRecommendations S3DestinationConfig

An object to specify the destination Amazon Simple Storage Service (Amazon S3) bucket name and key prefix for the export job.

You must create the destination Amazon S3 bucket for your recommendations export before you create the export job. Compute Optimizer does not create the S3 bucket for you. After you create the S3 bucket, ensure that it has the required permissions policy to allow Compute Optimizer to write the export file to it. If you plan to specify an object prefix when you create the export job, you must include the object prefix in the policy that you add to the S3 bucket. For more information, see Amazon S3 Bucket Policy for Compute Optimizer in the Compute Optimizer User Guide.

Destructuring the Response

data ExportAutoScalingGroupRecommendationsResponse

See: newExportAutoScalingGroupRecommendationsResponse smart constructor.

Constructors

ExportAutoScalingGroupRecommendationsResponse'

Fields

jobId :: Maybe Text

The identification number of the export job.

Use the DescribeRecommendationExportJobs action, and specify the job ID to view the status of an export job.

s3Destination :: Maybe S3Destination

An object that describes the destination Amazon S3 bucket of a recommendations export file.

httpStatus :: Int

The response's http status code.

newExportAutoScalingGroupRecommendationsResponse

Arguments

:: Int
$sel:httpStatus:ExportAutoScalingGroupRecommendationsResponse'

-> ExportAutoScalingGroupRecommendationsResponse

Create a value of ExportAutoScalingGroupRecommendationsResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:jobId:ExportAutoScalingGroupRecommendationsResponse', exportAutoScalingGroupRecommendationsResponse_jobId - The identification number of the export job.

Use the DescribeRecommendationExportJobs action, and specify the job ID to view the status of an export job.

$sel:s3Destination:ExportAutoScalingGroupRecommendationsResponse', exportAutoScalingGroupRecommendationsResponse_s3Destination - An object that describes the destination Amazon S3 bucket of a recommendations export file.

$sel:httpStatus:ExportAutoScalingGroupRecommendationsResponse', exportAutoScalingGroupRecommendationsResponse_httpStatus - The response's http status code.

Response Lenses

exportAutoScalingGroupRecommendationsResponse_jobId :: Lens' ExportAutoScalingGroupRecommendationsResponse (Maybe Text)

The identification number of the export job.

Use the DescribeRecommendationExportJobs action, and specify the job ID to view the status of an export job.

exportAutoScalingGroupRecommendationsResponse_s3Destination :: Lens' ExportAutoScalingGroupRecommendationsResponse (Maybe S3Destination)

An object that describes the destination Amazon S3 bucket of a recommendations export file.

exportAutoScalingGroupRecommendationsResponse_httpStatus :: Lens' ExportAutoScalingGroupRecommendationsResponse Int

The response's http status code.