Amazonka.ComputeOptimizer.Types.LambdaFunctionRecommendation

Description

Documentation

data LambdaFunctionRecommendation

Describes an Lambda function recommendation.

See: newLambdaFunctionRecommendation smart constructor.

Constructors

LambdaFunctionRecommendation'

Fields

functionArn :: Maybe Text

The Amazon Resource Name (ARN) of the current function.

finding :: Maybe LambdaFunctionRecommendationFinding

The finding classification of the function.

Findings for functions include:

Optimized — The function is correctly provisioned to run your workload based on its current configuration and its utilization history. This finding classification does not include finding reason codes.

NotOptimized — The function is performing at a higher level (over-provisioned) or at a lower level (under-provisioned) than required for your workload because its current configuration is not optimal. Over-provisioned resources might lead to unnecessary infrastructure cost, and under-provisioned resources might lead to poor application performance. This finding classification can include the MemoryUnderprovisioned and MemoryUnderprovisioned finding reason codes.

Unavailable — Compute Optimizer was unable to generate a recommendation for the function. This could be because the function has not accumulated sufficient metric data, or the function does not qualify for a recommendation. This finding classification can include the InsufficientData and Inconclusive finding reason codes.

Functions with a finding of unavailable are not returned unless you specify the filter parameter with a value of Unavailable in your GetLambdaFunctionRecommendations request.

currentMemorySize :: Maybe Int

The amount of memory, in MB, that's allocated to the current function.

lastRefreshTimestamp :: Maybe POSIX

The timestamp of when the function recommendation was last refreshed.

accountId :: Maybe Text

The Amazon Web Services account ID of the function.

findingReasonCodes :: Maybe [LambdaFunctionRecommendationFindingReasonCode]

The reason for the finding classification of the function.

Functions that have a finding classification of Optimized don't have a finding reason code.

Finding reason codes for functions include:

MemoryOverprovisioned — The function is over-provisioned when its memory configuration can be sized down while still meeting the performance requirements of your workload. An over-provisioned function might lead to unnecessary infrastructure cost. This finding reason code is part of the NotOptimized finding classification.

MemoryUnderprovisioned — The function is under-provisioned when its memory configuration doesn't meet the performance requirements of the workload. An under-provisioned function might lead to poor application performance. This finding reason code is part of the NotOptimized finding classification.

InsufficientData — The function does not have sufficient metric data for Compute Optimizer to generate a recommendation. For more information, see the Supported resources and requirements in the Compute Optimizer User Guide. This finding reason code is part of the Unavailable finding classification.

Inconclusive — The function does not qualify for a recommendation because Compute Optimizer cannot generate a recommendation with a high degree of confidence. This finding reason code is part of the Unavailable finding classification.

utilizationMetrics :: Maybe [LambdaFunctionUtilizationMetric]

An array of objects that describe the utilization metrics of the function.

memorySizeRecommendationOptions :: Maybe [LambdaFunctionMemoryRecommendationOption]

An array of objects that describe the memory configuration recommendation options for the function.

functionVersion :: Maybe Text

The version number of the current function.

numberOfInvocations :: Maybe Integer

The number of times your function code was applied during the look-back period.

lookbackPeriodInDays :: Maybe Double

The number of days for which utilization metrics were analyzed for the function.

newLambdaFunctionRecommendation :: LambdaFunctionRecommendation

Create a value of LambdaFunctionRecommendation with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:functionArn:LambdaFunctionRecommendation', lambdaFunctionRecommendation_functionArn - The Amazon Resource Name (ARN) of the current function.

$sel:finding:LambdaFunctionRecommendation', lambdaFunctionRecommendation_finding - The finding classification of the function.

Findings for functions include:

Optimized — The function is correctly provisioned to run your workload based on its current configuration and its utilization history. This finding classification does not include finding reason codes.

NotOptimized — The function is performing at a higher level (over-provisioned) or at a lower level (under-provisioned) than required for your workload because its current configuration is not optimal. Over-provisioned resources might lead to unnecessary infrastructure cost, and under-provisioned resources might lead to poor application performance. This finding classification can include the MemoryUnderprovisioned and MemoryUnderprovisioned finding reason codes.

Unavailable — Compute Optimizer was unable to generate a recommendation for the function. This could be because the function has not accumulated sufficient metric data, or the function does not qualify for a recommendation. This finding classification can include the InsufficientData and Inconclusive finding reason codes.

Functions with a finding of unavailable are not returned unless you specify the filter parameter with a value of Unavailable in your GetLambdaFunctionRecommendations request.

$sel:currentMemorySize:LambdaFunctionRecommendation', lambdaFunctionRecommendation_currentMemorySize - The amount of memory, in MB, that's allocated to the current function.

$sel:lastRefreshTimestamp:LambdaFunctionRecommendation', lambdaFunctionRecommendation_lastRefreshTimestamp - The timestamp of when the function recommendation was last refreshed.

$sel:accountId:LambdaFunctionRecommendation', lambdaFunctionRecommendation_accountId - The Amazon Web Services account ID of the function.

$sel:findingReasonCodes:LambdaFunctionRecommendation', lambdaFunctionRecommendation_findingReasonCodes - The reason for the finding classification of the function.

Functions that have a finding classification of Optimized don't have a finding reason code.

Finding reason codes for functions include:

MemoryOverprovisioned — The function is over-provisioned when its memory configuration can be sized down while still meeting the performance requirements of your workload. An over-provisioned function might lead to unnecessary infrastructure cost. This finding reason code is part of the NotOptimized finding classification.

MemoryUnderprovisioned — The function is under-provisioned when its memory configuration doesn't meet the performance requirements of the workload. An under-provisioned function might lead to poor application performance. This finding reason code is part of the NotOptimized finding classification.

InsufficientData — The function does not have sufficient metric data for Compute Optimizer to generate a recommendation. For more information, see the Supported resources and requirements in the Compute Optimizer User Guide. This finding reason code is part of the Unavailable finding classification.

Inconclusive — The function does not qualify for a recommendation because Compute Optimizer cannot generate a recommendation with a high degree of confidence. This finding reason code is part of the Unavailable finding classification.

$sel:utilizationMetrics:LambdaFunctionRecommendation', lambdaFunctionRecommendation_utilizationMetrics - An array of objects that describe the utilization metrics of the function.

$sel:memorySizeRecommendationOptions:LambdaFunctionRecommendation', lambdaFunctionRecommendation_memorySizeRecommendationOptions - An array of objects that describe the memory configuration recommendation options for the function.

$sel:functionVersion:LambdaFunctionRecommendation', lambdaFunctionRecommendation_functionVersion - The version number of the current function.

$sel:numberOfInvocations:LambdaFunctionRecommendation', lambdaFunctionRecommendation_numberOfInvocations - The number of times your function code was applied during the look-back period.

$sel:lookbackPeriodInDays:LambdaFunctionRecommendation', lambdaFunctionRecommendation_lookbackPeriodInDays - The number of days for which utilization metrics were analyzed for the function.

lambdaFunctionRecommendation_functionArn :: Lens' LambdaFunctionRecommendation (Maybe Text)

The Amazon Resource Name (ARN) of the current function.

lambdaFunctionRecommendation_finding :: Lens' LambdaFunctionRecommendation (Maybe LambdaFunctionRecommendationFinding)

The finding classification of the function.

Findings for functions include:

Optimized — The function is correctly provisioned to run your workload based on its current configuration and its utilization history. This finding classification does not include finding reason codes.

NotOptimized — The function is performing at a higher level (over-provisioned) or at a lower level (under-provisioned) than required for your workload because its current configuration is not optimal. Over-provisioned resources might lead to unnecessary infrastructure cost, and under-provisioned resources might lead to poor application performance. This finding classification can include the MemoryUnderprovisioned and MemoryUnderprovisioned finding reason codes.

Unavailable — Compute Optimizer was unable to generate a recommendation for the function. This could be because the function has not accumulated sufficient metric data, or the function does not qualify for a recommendation. This finding classification can include the InsufficientData and Inconclusive finding reason codes.

Functions with a finding of unavailable are not returned unless you specify the filter parameter with a value of Unavailable in your GetLambdaFunctionRecommendations request.

lambdaFunctionRecommendation_currentMemorySize :: Lens' LambdaFunctionRecommendation (Maybe Int)

The amount of memory, in MB, that's allocated to the current function.

lambdaFunctionRecommendation_lastRefreshTimestamp :: Lens' LambdaFunctionRecommendation (Maybe UTCTime)

The timestamp of when the function recommendation was last refreshed.

lambdaFunctionRecommendation_accountId :: Lens' LambdaFunctionRecommendation (Maybe Text)

The Amazon Web Services account ID of the function.

lambdaFunctionRecommendation_findingReasonCodes :: Lens' LambdaFunctionRecommendation (Maybe [LambdaFunctionRecommendationFindingReasonCode])

The reason for the finding classification of the function.

Functions that have a finding classification of Optimized don't have a finding reason code.

Finding reason codes for functions include:

MemoryOverprovisioned — The function is over-provisioned when its memory configuration can be sized down while still meeting the performance requirements of your workload. An over-provisioned function might lead to unnecessary infrastructure cost. This finding reason code is part of the NotOptimized finding classification.

MemoryUnderprovisioned — The function is under-provisioned when its memory configuration doesn't meet the performance requirements of the workload. An under-provisioned function might lead to poor application performance. This finding reason code is part of the NotOptimized finding classification.

InsufficientData — The function does not have sufficient metric data for Compute Optimizer to generate a recommendation. For more information, see the Supported resources and requirements in the Compute Optimizer User Guide. This finding reason code is part of the Unavailable finding classification.

Inconclusive — The function does not qualify for a recommendation because Compute Optimizer cannot generate a recommendation with a high degree of confidence. This finding reason code is part of the Unavailable finding classification.

lambdaFunctionRecommendation_utilizationMetrics :: Lens' LambdaFunctionRecommendation (Maybe [LambdaFunctionUtilizationMetric])

An array of objects that describe the utilization metrics of the function.

lambdaFunctionRecommendation_memorySizeRecommendationOptions :: Lens' LambdaFunctionRecommendation (Maybe [LambdaFunctionMemoryRecommendationOption])

An array of objects that describe the memory configuration recommendation options for the function.

lambdaFunctionRecommendation_functionVersion :: Lens' LambdaFunctionRecommendation (Maybe Text)

The version number of the current function.

lambdaFunctionRecommendation_numberOfInvocations :: Lens' LambdaFunctionRecommendation (Maybe Integer)

The number of times your function code was applied during the look-back period.

lambdaFunctionRecommendation_lookbackPeriodInDays :: Lens' LambdaFunctionRecommendation (Maybe Double)

The number of days for which utilization metrics were analyzed for the function.