Amazonka.ComputeOptimizer.Types.RecommendationSummary

Description

Documentation

data RecommendationSummary

A summary of a recommendation.

See: newRecommendationSummary smart constructor.

Constructors

RecommendationSummary'

Fields

accountId :: Maybe Text

The Amazon Web Services account ID of the recommendation summary.

summaries :: Maybe [Summary]

An array of objects that describe a recommendation summary.

recommendationResourceType :: Maybe RecommendationSourceType

The resource type of the recommendation.

newRecommendationSummary :: RecommendationSummary

Create a value of RecommendationSummary with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:accountId:RecommendationSummary', recommendationSummary_accountId - The Amazon Web Services account ID of the recommendation summary.

$sel:summaries:RecommendationSummary', recommendationSummary_summaries - An array of objects that describe a recommendation summary.

$sel:recommendationResourceType:RecommendationSummary', recommendationSummary_recommendationResourceType - The resource type of the recommendation.

recommendationSummary_accountId :: Lens' RecommendationSummary (Maybe Text)

The Amazon Web Services account ID of the recommendation summary.

recommendationSummary_summaries :: Lens' RecommendationSummary (Maybe [Summary])

An array of objects that describe a recommendation summary.

recommendationSummary_recommendationResourceType :: Lens' RecommendationSummary (Maybe RecommendationSourceType)

The resource type of the recommendation.