Amazonka.Config.DeleteStoredQuery

Description

Deletes the stored query for a single Amazon Web Services account and a single Amazon Web Services Region.

Creating a Request

data DeleteStoredQuery

See: newDeleteStoredQuery smart constructor.

Constructors

DeleteStoredQuery'

Fields

queryName :: Text

The name of the query that you want to delete.

newDeleteStoredQuery

Arguments

:: Text
$sel:queryName:DeleteStoredQuery'

-> DeleteStoredQuery

Create a value of DeleteStoredQuery with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:queryName:DeleteStoredQuery', deleteStoredQuery_queryName - The name of the query that you want to delete.

Request Lenses

deleteStoredQuery_queryName :: Lens' DeleteStoredQuery Text

The name of the query that you want to delete.

Destructuring the Response

data DeleteStoredQueryResponse

See: newDeleteStoredQueryResponse smart constructor.

Constructors

DeleteStoredQueryResponse'

Fields

httpStatus :: Int

The response's http status code.

newDeleteStoredQueryResponse

Arguments

:: Int
$sel:httpStatus:DeleteStoredQueryResponse'

-> DeleteStoredQueryResponse

Create a value of DeleteStoredQueryResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:httpStatus:DeleteStoredQueryResponse', deleteStoredQueryResponse_httpStatus - The response's http status code.

Response Lenses

deleteStoredQueryResponse_httpStatus :: Lens' DeleteStoredQueryResponse Int

The response's http status code.