Amazonka.Config.DescribeConfigRuleEvaluationStatus

Description

Returns status information for each of your Config managed rules. The status includes information such as the last time Config invoked the rule, the last time Config failed to invoke the rule, and the related error for the last failure.

This operation returns paginated results.

Creating a Request

data DescribeConfigRuleEvaluationStatus

See: newDescribeConfigRuleEvaluationStatus smart constructor.

Constructors

DescribeConfigRuleEvaluationStatus'

Fields

configRuleNames :: Maybe [Text]

The name of the Config managed rules for which you want status information. If you do not specify any names, Config returns status information for all Config managed rules that you use.

nextToken :: Maybe Text

The nextToken string returned on a previous page that you use to get the next page of results in a paginated response.

limit :: Maybe Natural

The number of rule evaluation results that you want returned.

This parameter is required if the rule limit for your account is more than the default of 150 rules.

For information about requesting a rule limit increase, see Config Limits in the Amazon Web Services General Reference Guide.

newDescribeConfigRuleEvaluationStatus :: DescribeConfigRuleEvaluationStatus

Create a value of DescribeConfigRuleEvaluationStatus with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:configRuleNames:DescribeConfigRuleEvaluationStatus', describeConfigRuleEvaluationStatus_configRuleNames - The name of the Config managed rules for which you want status information. If you do not specify any names, Config returns status information for all Config managed rules that you use.

$sel:nextToken:DescribeConfigRuleEvaluationStatus', describeConfigRuleEvaluationStatus_nextToken - The nextToken string returned on a previous page that you use to get the next page of results in a paginated response.

$sel:limit:DescribeConfigRuleEvaluationStatus', describeConfigRuleEvaluationStatus_limit - The number of rule evaluation results that you want returned.

This parameter is required if the rule limit for your account is more than the default of 150 rules.

For information about requesting a rule limit increase, see Config Limits in the Amazon Web Services General Reference Guide.

Request Lenses

describeConfigRuleEvaluationStatus_configRuleNames :: Lens' DescribeConfigRuleEvaluationStatus (Maybe [Text])

The name of the Config managed rules for which you want status information. If you do not specify any names, Config returns status information for all Config managed rules that you use.

describeConfigRuleEvaluationStatus_nextToken :: Lens' DescribeConfigRuleEvaluationStatus (Maybe Text)

The nextToken string returned on a previous page that you use to get the next page of results in a paginated response.

describeConfigRuleEvaluationStatus_limit :: Lens' DescribeConfigRuleEvaluationStatus (Maybe Natural)

The number of rule evaluation results that you want returned.

This parameter is required if the rule limit for your account is more than the default of 150 rules.

For information about requesting a rule limit increase, see Config Limits in the Amazon Web Services General Reference Guide.

Destructuring the Response

data DescribeConfigRuleEvaluationStatusResponse

See: newDescribeConfigRuleEvaluationStatusResponse smart constructor.

Constructors

DescribeConfigRuleEvaluationStatusResponse'

Fields

configRulesEvaluationStatus :: Maybe [ConfigRuleEvaluationStatus]

Status information about your Config managed rules.

nextToken :: Maybe Text

The string that you use in a subsequent request to get the next page of results in a paginated response.

httpStatus :: Int

The response's http status code.

newDescribeConfigRuleEvaluationStatusResponse

Arguments

:: Int
$sel:httpStatus:DescribeConfigRuleEvaluationStatusResponse'

-> DescribeConfigRuleEvaluationStatusResponse

Create a value of DescribeConfigRuleEvaluationStatusResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:configRulesEvaluationStatus:DescribeConfigRuleEvaluationStatusResponse', describeConfigRuleEvaluationStatusResponse_configRulesEvaluationStatus - Status information about your Config managed rules.

$sel:nextToken:DescribeConfigRuleEvaluationStatus', describeConfigRuleEvaluationStatusResponse_nextToken - The string that you use in a subsequent request to get the next page of results in a paginated response.

$sel:httpStatus:DescribeConfigRuleEvaluationStatusResponse', describeConfigRuleEvaluationStatusResponse_httpStatus - The response's http status code.

Response Lenses

describeConfigRuleEvaluationStatusResponse_configRulesEvaluationStatus :: Lens' DescribeConfigRuleEvaluationStatusResponse (Maybe [ConfigRuleEvaluationStatus])

Status information about your Config managed rules.

describeConfigRuleEvaluationStatusResponse_nextToken :: Lens' DescribeConfigRuleEvaluationStatusResponse (Maybe Text)

The string that you use in a subsequent request to get the next page of results in a paginated response.

describeConfigRuleEvaluationStatusResponse_httpStatus :: Lens' DescribeConfigRuleEvaluationStatusResponse Int

The response's http status code.