Amazonka.Config.PutRemediationExceptions

Description

A remediation exception is when a specific resource is no longer considered for auto-remediation. This API adds a new exception or updates an existing exception for a specific resource with a specific Config rule.

Config generates a remediation exception when a problem occurs executing a remediation action to a specific resource. Remediation exceptions blocks auto-remediation until the exception is cleared.

Creating a Request

data PutRemediationExceptions

See: newPutRemediationExceptions smart constructor.

Constructors

PutRemediationExceptions'

Fields

message :: Maybe Text

The message contains an explanation of the exception.

expirationTime :: Maybe POSIX

The exception is automatically deleted after the expiration date.

configRuleName :: Text

The name of the Config rule for which you want to create remediation exception.

resourceKeys :: NonEmpty RemediationExceptionResourceKey

An exception list of resource exception keys to be processed with the current request. Config adds exception for each resource key. For example, Config adds 3 exceptions for 3 resource keys.

newPutRemediationExceptions

Arguments

:: Text
$sel:configRuleName:PutRemediationExceptions'

-> NonEmpty RemediationExceptionResourceKey
$sel:resourceKeys:PutRemediationExceptions'

-> PutRemediationExceptions

Create a value of PutRemediationExceptions with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:message:PutRemediationExceptions', putRemediationExceptions_message - The message contains an explanation of the exception.

$sel:expirationTime:PutRemediationExceptions', putRemediationExceptions_expirationTime - The exception is automatically deleted after the expiration date.

$sel:configRuleName:PutRemediationExceptions', putRemediationExceptions_configRuleName - The name of the Config rule for which you want to create remediation exception.

$sel:resourceKeys:PutRemediationExceptions', putRemediationExceptions_resourceKeys - An exception list of resource exception keys to be processed with the current request. Config adds exception for each resource key. For example, Config adds 3 exceptions for 3 resource keys.

Request Lenses

putRemediationExceptions_message :: Lens' PutRemediationExceptions (Maybe Text)

The message contains an explanation of the exception.

putRemediationExceptions_expirationTime :: Lens' PutRemediationExceptions (Maybe UTCTime)

The exception is automatically deleted after the expiration date.

putRemediationExceptions_configRuleName :: Lens' PutRemediationExceptions Text

The name of the Config rule for which you want to create remediation exception.

putRemediationExceptions_resourceKeys :: Lens' PutRemediationExceptions (NonEmpty RemediationExceptionResourceKey)

An exception list of resource exception keys to be processed with the current request. Config adds exception for each resource key. For example, Config adds 3 exceptions for 3 resource keys.

Destructuring the Response

data PutRemediationExceptionsResponse

See: newPutRemediationExceptionsResponse smart constructor.

Constructors

PutRemediationExceptionsResponse'

Fields

failedBatches :: Maybe [FailedRemediationExceptionBatch]

Returns a list of failed remediation exceptions batch objects. Each object in the batch consists of a list of failed items and failure messages.

httpStatus :: Int

The response's http status code.

newPutRemediationExceptionsResponse

Arguments

:: Int
$sel:httpStatus:PutRemediationExceptionsResponse'

-> PutRemediationExceptionsResponse

Create a value of PutRemediationExceptionsResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:failedBatches:PutRemediationExceptionsResponse', putRemediationExceptionsResponse_failedBatches - Returns a list of failed remediation exceptions batch objects. Each object in the batch consists of a list of failed items and failure messages.

$sel:httpStatus:PutRemediationExceptionsResponse', putRemediationExceptionsResponse_httpStatus - The response's http status code.

Response Lenses

putRemediationExceptionsResponse_failedBatches :: Lens' PutRemediationExceptionsResponse (Maybe [FailedRemediationExceptionBatch])

Returns a list of failed remediation exceptions batch objects. Each object in the batch consists of a list of failed items and failure messages.

putRemediationExceptionsResponse_httpStatus :: Lens' PutRemediationExceptionsResponse Int

The response's http status code.