Amazonka.Config.PutResourceConfig

Description

Records the configuration state for the resource provided in the request. The configuration state of a resource is represented in Config as Configuration Items. Once this API records the configuration item, you can retrieve the list of configuration items for the custom resource type using existing Config APIs.

The custom resource type must be registered with CloudFormation. This API accepts the configuration item registered with CloudFormation.

When you call this API, Config only stores configuration state of the resource provided in the request. This API does not change or remediate the configuration of the resource.

Write-only schema properites are not recorded as part of the published configuration item.

Creating a Request

data PutResourceConfig

See: newPutResourceConfig smart constructor.

Constructors

PutResourceConfig'

Fields

resourceName :: Maybe Text

Name of the resource.

tags :: Maybe (HashMap Text Text)

Tags associated with the resource.

resourceType :: Text

The type of the resource. The custom resource type must be registered with CloudFormation.

You cannot use the organization names “amzn”, “amazon”, “alexa”, “custom” with custom resource types. It is the first part of the ResourceType up to the first ::.

schemaVersionId :: Text

Version of the schema registered for the ResourceType in CloudFormation.

resourceId :: Text

Unique identifier of the resource.

configuration :: Text

The configuration object of the resource in valid JSON format. It must match the schema registered with CloudFormation.

The configuration JSON must not exceed 64 KB.

newPutResourceConfig

Arguments

:: Text
$sel:resourceType:PutResourceConfig'

-> Text
$sel:schemaVersionId:PutResourceConfig'

-> Text
$sel:resourceId:PutResourceConfig'

-> Text
$sel:configuration:PutResourceConfig'

-> PutResourceConfig

Create a value of PutResourceConfig with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:resourceName:PutResourceConfig', putResourceConfig_resourceName - Name of the resource.

$sel:tags:PutResourceConfig', putResourceConfig_tags - Tags associated with the resource.

$sel:resourceType:PutResourceConfig', putResourceConfig_resourceType - The type of the resource. The custom resource type must be registered with CloudFormation.

You cannot use the organization names “amzn”, “amazon”, “alexa”, “custom” with custom resource types. It is the first part of the ResourceType up to the first ::.

$sel:schemaVersionId:PutResourceConfig', putResourceConfig_schemaVersionId - Version of the schema registered for the ResourceType in CloudFormation.

$sel:resourceId:PutResourceConfig', putResourceConfig_resourceId - Unique identifier of the resource.

$sel:configuration:PutResourceConfig', putResourceConfig_configuration - The configuration object of the resource in valid JSON format. It must match the schema registered with CloudFormation.

The configuration JSON must not exceed 64 KB.

Request Lenses

putResourceConfig_resourceName :: Lens' PutResourceConfig (Maybe Text)

Name of the resource.

putResourceConfig_tags :: Lens' PutResourceConfig (Maybe (HashMap Text Text))

Tags associated with the resource.

putResourceConfig_resourceType :: Lens' PutResourceConfig Text

The type of the resource. The custom resource type must be registered with CloudFormation.

You cannot use the organization names “amzn”, “amazon”, “alexa”, “custom” with custom resource types. It is the first part of the ResourceType up to the first ::.

putResourceConfig_schemaVersionId :: Lens' PutResourceConfig Text

Version of the schema registered for the ResourceType in CloudFormation.

putResourceConfig_resourceId :: Lens' PutResourceConfig Text

Unique identifier of the resource.

putResourceConfig_configuration :: Lens' PutResourceConfig Text

The configuration object of the resource in valid JSON format. It must match the schema registered with CloudFormation.

The configuration JSON must not exceed 64 KB.

Destructuring the Response

data PutResourceConfigResponse

See: newPutResourceConfigResponse smart constructor.

Constructors

PutResourceConfigResponse'

newPutResourceConfigResponse :: PutResourceConfigResponse

Create a value of PutResourceConfigResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.