Amazonka.Config.Types.ComplianceByConfigRule

Description

Documentation

data ComplianceByConfigRule

Indicates whether an Config rule is compliant. A rule is compliant if all of the resources that the rule evaluated comply with it. A rule is noncompliant if any of these resources do not comply.

See: newComplianceByConfigRule smart constructor.

Constructors

ComplianceByConfigRule'

Fields

compliance :: Maybe Compliance

Indicates whether the Config rule is compliant.

configRuleName :: Maybe Text

The name of the Config rule.

newComplianceByConfigRule :: ComplianceByConfigRule

Create a value of ComplianceByConfigRule with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:compliance:ComplianceByConfigRule', complianceByConfigRule_compliance - Indicates whether the Config rule is compliant.

$sel:configRuleName:ComplianceByConfigRule', complianceByConfigRule_configRuleName - The name of the Config rule.

complianceByConfigRule_compliance :: Lens' ComplianceByConfigRule (Maybe Compliance)

Indicates whether the Config rule is compliant.

complianceByConfigRule_configRuleName :: Lens' ComplianceByConfigRule (Maybe Text)

The name of the Config rule.