Amazonka.Config.Types.ConfigurationAggregator

Description

Documentation

data ConfigurationAggregator

The details about the configuration aggregator, including information about source accounts, regions, and metadata of the aggregator.

See: newConfigurationAggregator smart constructor.

Constructors

ConfigurationAggregator'

Fields

configurationAggregatorArn :: Maybe Text

The Amazon Resource Name (ARN) of the aggregator.

creationTime :: Maybe POSIX

The time stamp when the configuration aggregator was created.

organizationAggregationSource :: Maybe OrganizationAggregationSource

Provides an organization and list of regions to be aggregated.

lastUpdatedTime :: Maybe POSIX

The time of the last update.

accountAggregationSources :: Maybe [AccountAggregationSource]

Provides a list of source accounts and regions to be aggregated.

createdBy :: Maybe Text

Amazon Web Services service that created the configuration aggregator.

configurationAggregatorName :: Maybe Text

The name of the aggregator.

newConfigurationAggregator :: ConfigurationAggregator

Create a value of ConfigurationAggregator with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:configurationAggregatorArn:ConfigurationAggregator', configurationAggregator_configurationAggregatorArn - The Amazon Resource Name (ARN) of the aggregator.

$sel:creationTime:ConfigurationAggregator', configurationAggregator_creationTime - The time stamp when the configuration aggregator was created.

$sel:organizationAggregationSource:ConfigurationAggregator', configurationAggregator_organizationAggregationSource - Provides an organization and list of regions to be aggregated.

$sel:lastUpdatedTime:ConfigurationAggregator', configurationAggregator_lastUpdatedTime - The time of the last update.

$sel:accountAggregationSources:ConfigurationAggregator', configurationAggregator_accountAggregationSources - Provides a list of source accounts and regions to be aggregated.

$sel:createdBy:ConfigurationAggregator', configurationAggregator_createdBy - Amazon Web Services service that created the configuration aggregator.

$sel:configurationAggregatorName:ConfigurationAggregator', configurationAggregator_configurationAggregatorName - The name of the aggregator.

configurationAggregator_configurationAggregatorArn :: Lens' ConfigurationAggregator (Maybe Text)

The Amazon Resource Name (ARN) of the aggregator.

configurationAggregator_creationTime :: Lens' ConfigurationAggregator (Maybe UTCTime)

The time stamp when the configuration aggregator was created.

configurationAggregator_organizationAggregationSource :: Lens' ConfigurationAggregator (Maybe OrganizationAggregationSource)

Provides an organization and list of regions to be aggregated.

configurationAggregator_lastUpdatedTime :: Lens' ConfigurationAggregator (Maybe UTCTime)

The time of the last update.

configurationAggregator_accountAggregationSources :: Lens' ConfigurationAggregator (Maybe [AccountAggregationSource])

Provides a list of source accounts and regions to be aggregated.

configurationAggregator_createdBy :: Lens' ConfigurationAggregator (Maybe Text)

Amazon Web Services service that created the configuration aggregator.

configurationAggregator_configurationAggregatorName :: Lens' ConfigurationAggregator (Maybe Text)

The name of the aggregator.