Amazonka.Config.Types.ConfigurationItem

Description

Documentation

data ConfigurationItem

A list that contains detailed configurations of a specified resource.

See: newConfigurationItem smart constructor.

Constructors

ConfigurationItem'

Fields

resourceId :: Maybe Text

The ID of the resource (for example, sg-xxxxxx).

resourceType :: Maybe ResourceType

The type of Amazon Web Services resource.

configurationStateId :: Maybe Text

An identifier that indicates the ordering of the configuration items of a resource.

arn :: Maybe Text

Amazon Resource Name (ARN) associated with the resource.

resourceName :: Maybe Text

The custom name of the resource, if available.

resourceCreationTime :: Maybe POSIX

The time stamp when the resource was created.

configurationItemStatus :: Maybe ConfigurationItemStatus

The configuration item status. The valid values are:

OK – The resource configuration has been updated

ResourceDiscovered – The resource was newly discovered

ResourceNotRecorded – The resource was discovered but its configuration was not recorded since the recorder excludes the recording of resources of this type

ResourceDeleted – The resource was deleted

ResourceDeletedNotRecorded – The resource was deleted but its configuration was not recorded since the recorder excludes the recording of resources of this type

The CIs do not incur any cost.

configurationItemCaptureTime :: Maybe POSIX

The time when the configuration recording was initiated.

accountId :: Maybe Text

The 12-digit Amazon Web Services account ID associated with the resource.

supplementaryConfiguration :: Maybe (HashMap Text Text)

Configuration attributes that Config returns for certain resource types to supplement the information returned for the configuration parameter.

availabilityZone :: Maybe Text

The Availability Zone associated with the resource.

relationships :: Maybe [Relationship]

A list of related Amazon Web Services resources.

version :: Maybe Text

The version number of the resource configuration.

awsRegion :: Maybe Text

The region where the resource resides.

relatedEvents :: Maybe [Text]

A list of CloudTrail event IDs.

A populated field indicates that the current configuration was initiated by the events recorded in the CloudTrail log. For more information about CloudTrail, see What Is CloudTrail.

An empty field indicates that the current configuration was not initiated by any event. As of Version 1.3, the relatedEvents field is empty. You can access the LookupEvents API in the CloudTrail API Reference to retrieve the events for the resource.

configuration :: Maybe Text

The description of the resource configuration.

configurationItemMD5Hash :: Maybe Text

Unique MD5 hash that represents the configuration item's state.

You can use MD5 hash to compare the states of two or more configuration items that are associated with the same resource.

tags :: Maybe (HashMap Text Text)

A mapping of key value tags associated with the resource.

newConfigurationItem :: ConfigurationItem

Create a value of ConfigurationItem with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:resourceId:ConfigurationItem', configurationItem_resourceId - The ID of the resource (for example, sg-xxxxxx).

$sel:resourceType:ConfigurationItem', configurationItem_resourceType - The type of Amazon Web Services resource.

$sel:configurationStateId:ConfigurationItem', configurationItem_configurationStateId - An identifier that indicates the ordering of the configuration items of a resource.

$sel:arn:ConfigurationItem', configurationItem_arn - Amazon Resource Name (ARN) associated with the resource.

$sel:resourceName:ConfigurationItem', configurationItem_resourceName - The custom name of the resource, if available.

$sel:resourceCreationTime:ConfigurationItem', configurationItem_resourceCreationTime - The time stamp when the resource was created.

$sel:configurationItemStatus:ConfigurationItem', configurationItem_configurationItemStatus - The configuration item status. The valid values are:

OK – The resource configuration has been updated

ResourceDiscovered – The resource was newly discovered

ResourceNotRecorded – The resource was discovered but its configuration was not recorded since the recorder excludes the recording of resources of this type

ResourceDeleted – The resource was deleted

ResourceDeletedNotRecorded – The resource was deleted but its configuration was not recorded since the recorder excludes the recording of resources of this type

The CIs do not incur any cost.

$sel:configurationItemCaptureTime:ConfigurationItem', configurationItem_configurationItemCaptureTime - The time when the configuration recording was initiated.

$sel:accountId:ConfigurationItem', configurationItem_accountId - The 12-digit Amazon Web Services account ID associated with the resource.

$sel:supplementaryConfiguration:ConfigurationItem', configurationItem_supplementaryConfiguration - Configuration attributes that Config returns for certain resource types to supplement the information returned for the configuration parameter.

$sel:availabilityZone:ConfigurationItem', configurationItem_availabilityZone - The Availability Zone associated with the resource.

$sel:relationships:ConfigurationItem', configurationItem_relationships - A list of related Amazon Web Services resources.

$sel:version:ConfigurationItem', configurationItem_version - The version number of the resource configuration.

$sel:awsRegion:ConfigurationItem', configurationItem_awsRegion - The region where the resource resides.

$sel:relatedEvents:ConfigurationItem', configurationItem_relatedEvents - A list of CloudTrail event IDs.

A populated field indicates that the current configuration was initiated by the events recorded in the CloudTrail log. For more information about CloudTrail, see What Is CloudTrail.

An empty field indicates that the current configuration was not initiated by any event. As of Version 1.3, the relatedEvents field is empty. You can access the LookupEvents API in the CloudTrail API Reference to retrieve the events for the resource.

$sel:configuration:ConfigurationItem', configurationItem_configuration - The description of the resource configuration.

$sel:configurationItemMD5Hash:ConfigurationItem', configurationItem_configurationItemMD5Hash - Unique MD5 hash that represents the configuration item's state.

You can use MD5 hash to compare the states of two or more configuration items that are associated with the same resource.

$sel:tags:ConfigurationItem', configurationItem_tags - A mapping of key value tags associated with the resource.

configurationItem_resourceId :: Lens' ConfigurationItem (Maybe Text)

The ID of the resource (for example, sg-xxxxxx).

configurationItem_resourceType :: Lens' ConfigurationItem (Maybe ResourceType)

The type of Amazon Web Services resource.

configurationItem_configurationStateId :: Lens' ConfigurationItem (Maybe Text)

An identifier that indicates the ordering of the configuration items of a resource.

configurationItem_arn :: Lens' ConfigurationItem (Maybe Text)

Amazon Resource Name (ARN) associated with the resource.

configurationItem_resourceName :: Lens' ConfigurationItem (Maybe Text)

The custom name of the resource, if available.

configurationItem_resourceCreationTime :: Lens' ConfigurationItem (Maybe UTCTime)

The time stamp when the resource was created.

configurationItem_configurationItemStatus :: Lens' ConfigurationItem (Maybe ConfigurationItemStatus)

The configuration item status. The valid values are:

OK – The resource configuration has been updated

ResourceDiscovered – The resource was newly discovered

ResourceNotRecorded – The resource was discovered but its configuration was not recorded since the recorder excludes the recording of resources of this type

ResourceDeleted – The resource was deleted

ResourceDeletedNotRecorded – The resource was deleted but its configuration was not recorded since the recorder excludes the recording of resources of this type

The CIs do not incur any cost.

configurationItem_configurationItemCaptureTime :: Lens' ConfigurationItem (Maybe UTCTime)

The time when the configuration recording was initiated.

configurationItem_accountId :: Lens' ConfigurationItem (Maybe Text)

The 12-digit Amazon Web Services account ID associated with the resource.

configurationItem_supplementaryConfiguration :: Lens' ConfigurationItem (Maybe (HashMap Text Text))

Configuration attributes that Config returns for certain resource types to supplement the information returned for the configuration parameter.

configurationItem_availabilityZone :: Lens' ConfigurationItem (Maybe Text)

The Availability Zone associated with the resource.

configurationItem_relationships :: Lens' ConfigurationItem (Maybe [Relationship])

A list of related Amazon Web Services resources.

configurationItem_version :: Lens' ConfigurationItem (Maybe Text)

The version number of the resource configuration.

configurationItem_awsRegion :: Lens' ConfigurationItem (Maybe Text)

The region where the resource resides.

configurationItem_relatedEvents :: Lens' ConfigurationItem (Maybe [Text])

A list of CloudTrail event IDs.

A populated field indicates that the current configuration was initiated by the events recorded in the CloudTrail log. For more information about CloudTrail, see What Is CloudTrail.

An empty field indicates that the current configuration was not initiated by any event. As of Version 1.3, the relatedEvents field is empty. You can access the LookupEvents API in the CloudTrail API Reference to retrieve the events for the resource.

configurationItem_configuration :: Lens' ConfigurationItem (Maybe Text)

The description of the resource configuration.

configurationItem_configurationItemMD5Hash :: Lens' ConfigurationItem (Maybe Text)

Unique MD5 hash that represents the configuration item's state.

You can use MD5 hash to compare the states of two or more configuration items that are associated with the same resource.

configurationItem_tags :: Lens' ConfigurationItem (Maybe (HashMap Text Text))

A mapping of key value tags associated with the resource.