Amazonka.Config.Types.ConfigurationRecorderStatus

Description

Documentation

data ConfigurationRecorderStatus

The current status of the configuration recorder.

See: newConfigurationRecorderStatus smart constructor.

Constructors

ConfigurationRecorderStatus'

Fields

lastErrorCode :: Maybe Text

The error code indicating that the recording failed.

lastStopTime :: Maybe POSIX

The time the recorder was last stopped.

lastStatusChangeTime :: Maybe POSIX

The time when the status was last changed.

recording :: Maybe Bool

Specifies whether or not the recorder is currently recording.

lastStatus :: Maybe RecorderStatus

The last (previous) status of the recorder.

lastErrorMessage :: Maybe Text

The message indicating that the recording failed due to an error.

name :: Maybe Text

The name of the configuration recorder.

lastStartTime :: Maybe POSIX

The time the recorder was last started.

newConfigurationRecorderStatus :: ConfigurationRecorderStatus

Create a value of ConfigurationRecorderStatus with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:lastErrorCode:ConfigurationRecorderStatus', configurationRecorderStatus_lastErrorCode - The error code indicating that the recording failed.

$sel:lastStopTime:ConfigurationRecorderStatus', configurationRecorderStatus_lastStopTime - The time the recorder was last stopped.

$sel:lastStatusChangeTime:ConfigurationRecorderStatus', configurationRecorderStatus_lastStatusChangeTime - The time when the status was last changed.

$sel:recording:ConfigurationRecorderStatus', configurationRecorderStatus_recording - Specifies whether or not the recorder is currently recording.

$sel:lastStatus:ConfigurationRecorderStatus', configurationRecorderStatus_lastStatus - The last (previous) status of the recorder.

$sel:lastErrorMessage:ConfigurationRecorderStatus', configurationRecorderStatus_lastErrorMessage - The message indicating that the recording failed due to an error.

$sel:name:ConfigurationRecorderStatus', configurationRecorderStatus_name - The name of the configuration recorder.

$sel:lastStartTime:ConfigurationRecorderStatus', configurationRecorderStatus_lastStartTime - The time the recorder was last started.

configurationRecorderStatus_lastErrorCode :: Lens' ConfigurationRecorderStatus (Maybe Text)

The error code indicating that the recording failed.

configurationRecorderStatus_lastStopTime :: Lens' ConfigurationRecorderStatus (Maybe UTCTime)

The time the recorder was last stopped.

configurationRecorderStatus_lastStatusChangeTime :: Lens' ConfigurationRecorderStatus (Maybe UTCTime)

The time when the status was last changed.

configurationRecorderStatus_recording :: Lens' ConfigurationRecorderStatus (Maybe Bool)

Specifies whether or not the recorder is currently recording.

configurationRecorderStatus_lastStatus :: Lens' ConfigurationRecorderStatus (Maybe RecorderStatus)

The last (previous) status of the recorder.

configurationRecorderStatus_lastErrorMessage :: Lens' ConfigurationRecorderStatus (Maybe Text)

The message indicating that the recording failed due to an error.

configurationRecorderStatus_name :: Lens' ConfigurationRecorderStatus (Maybe Text)

The name of the configuration recorder.

configurationRecorderStatus_lastStartTime :: Lens' ConfigurationRecorderStatus (Maybe UTCTime)

The time the recorder was last started.