Amazonka.Config.Types.ConformancePackStatusDetail

Description

Documentation

data ConformancePackStatusDetail

Status details of a conformance pack.

See: newConformancePackStatusDetail smart constructor.

Constructors

ConformancePackStatusDetail'

Fields

conformancePackStatusReason :: Maybe Text

The reason of conformance pack creation failure.

lastUpdateCompletedTime :: Maybe POSIX

Last time when conformation pack creation and update was successful.

conformancePackName :: Text

Name of the conformance pack.

conformancePackId :: Text

ID of the conformance pack.

conformancePackArn :: Text

Amazon Resource Name (ARN) of comformance pack.

conformancePackState :: ConformancePackState

Indicates deployment status of conformance pack.

Config sets the state of the conformance pack to:

CREATE_IN_PROGRESS when a conformance pack creation is in progress for an account.

CREATE_COMPLETE when a conformance pack has been successfully created in your account.

CREATE_FAILED when a conformance pack creation failed in your account.

DELETE_IN_PROGRESS when a conformance pack deletion is in progress.

DELETE_FAILED when a conformance pack deletion failed in your account.

stackArn :: Text

Amazon Resource Name (ARN) of CloudFormation stack.

lastUpdateRequestedTime :: POSIX

Last time when conformation pack creation and update was requested.

newConformancePackStatusDetail

Arguments

:: Text
$sel:conformancePackName:ConformancePackStatusDetail'

-> Text
$sel:conformancePackId:ConformancePackStatusDetail'

-> Text
$sel:conformancePackArn:ConformancePackStatusDetail'

-> ConformancePackState
$sel:conformancePackState:ConformancePackStatusDetail'

-> Text
$sel:stackArn:ConformancePackStatusDetail'

-> UTCTime
$sel:lastUpdateRequestedTime:ConformancePackStatusDetail'

-> ConformancePackStatusDetail

Create a value of ConformancePackStatusDetail with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:conformancePackStatusReason:ConformancePackStatusDetail', conformancePackStatusDetail_conformancePackStatusReason - The reason of conformance pack creation failure.

$sel:lastUpdateCompletedTime:ConformancePackStatusDetail', conformancePackStatusDetail_lastUpdateCompletedTime - Last time when conformation pack creation and update was successful.

$sel:conformancePackName:ConformancePackStatusDetail', conformancePackStatusDetail_conformancePackName - Name of the conformance pack.

$sel:conformancePackId:ConformancePackStatusDetail', conformancePackStatusDetail_conformancePackId - ID of the conformance pack.

$sel:conformancePackArn:ConformancePackStatusDetail', conformancePackStatusDetail_conformancePackArn - Amazon Resource Name (ARN) of comformance pack.

$sel:conformancePackState:ConformancePackStatusDetail', conformancePackStatusDetail_conformancePackState - Indicates deployment status of conformance pack.

Config sets the state of the conformance pack to:

CREATE_IN_PROGRESS when a conformance pack creation is in progress for an account.

CREATE_COMPLETE when a conformance pack has been successfully created in your account.

CREATE_FAILED when a conformance pack creation failed in your account.

DELETE_IN_PROGRESS when a conformance pack deletion is in progress.

DELETE_FAILED when a conformance pack deletion failed in your account.

$sel:stackArn:ConformancePackStatusDetail', conformancePackStatusDetail_stackArn - Amazon Resource Name (ARN) of CloudFormation stack.

$sel:lastUpdateRequestedTime:ConformancePackStatusDetail', conformancePackStatusDetail_lastUpdateRequestedTime - Last time when conformation pack creation and update was requested.

conformancePackStatusDetail_conformancePackStatusReason :: Lens' ConformancePackStatusDetail (Maybe Text)

The reason of conformance pack creation failure.

conformancePackStatusDetail_lastUpdateCompletedTime :: Lens' ConformancePackStatusDetail (Maybe UTCTime)

Last time when conformation pack creation and update was successful.

conformancePackStatusDetail_conformancePackName :: Lens' ConformancePackStatusDetail Text

Name of the conformance pack.

conformancePackStatusDetail_conformancePackId :: Lens' ConformancePackStatusDetail Text

ID of the conformance pack.

conformancePackStatusDetail_conformancePackArn :: Lens' ConformancePackStatusDetail Text

Amazon Resource Name (ARN) of comformance pack.

conformancePackStatusDetail_conformancePackState :: Lens' ConformancePackStatusDetail ConformancePackState

Indicates deployment status of conformance pack.

Config sets the state of the conformance pack to:

CREATE_IN_PROGRESS when a conformance pack creation is in progress for an account.

CREATE_COMPLETE when a conformance pack has been successfully created in your account.

CREATE_FAILED when a conformance pack creation failed in your account.

DELETE_IN_PROGRESS when a conformance pack deletion is in progress.

DELETE_FAILED when a conformance pack deletion failed in your account.

conformancePackStatusDetail_stackArn :: Lens' ConformancePackStatusDetail Text

Amazon Resource Name (ARN) of CloudFormation stack.

conformancePackStatusDetail_lastUpdateRequestedTime :: Lens' ConformancePackStatusDetail UTCTime

Last time when conformation pack creation and update was requested.