Amazonka.Config.Types.Evaluation

Description

Documentation

data Evaluation

Identifies an Amazon Web Services resource and indicates whether it complies with the Config rule that it was evaluated against.

See: newEvaluation smart constructor.

Constructors

Evaluation'

Fields

annotation :: Maybe Text

Supplementary information about how the evaluation determined the compliance.

complianceResourceType :: Text

The type of Amazon Web Services resource that was evaluated.

complianceResourceId :: Text

The ID of the Amazon Web Services resource that was evaluated.

complianceType :: ComplianceType

Indicates whether the Amazon Web Services resource complies with the Config rule that it was evaluated against.

For the Evaluation data type, Config supports only the COMPLIANT, NON_COMPLIANT, and NOT_APPLICABLE values. Config does not support the INSUFFICIENT_DATA value for this data type.

Similarly, Config does not accept INSUFFICIENT_DATA as the value for ComplianceType from a PutEvaluations request. For example, an Lambda function for a custom Config rule cannot pass an INSUFFICIENT_DATA value to Config.

orderingTimestamp :: POSIX

The time of the event in Config that triggered the evaluation. For event-based evaluations, the time indicates when Config created the configuration item that triggered the evaluation. For periodic evaluations, the time indicates when Config triggered the evaluation at the frequency that you specified (for example, every 24 hours).

newEvaluation

Arguments

:: Text
$sel:complianceResourceType:Evaluation'

-> Text
$sel:complianceResourceId:Evaluation'

-> ComplianceType
$sel:complianceType:Evaluation'

-> UTCTime
$sel:orderingTimestamp:Evaluation'

-> Evaluation

Create a value of Evaluation with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:annotation:Evaluation', evaluation_annotation - Supplementary information about how the evaluation determined the compliance.

$sel:complianceResourceType:Evaluation', evaluation_complianceResourceType - The type of Amazon Web Services resource that was evaluated.

$sel:complianceResourceId:Evaluation', evaluation_complianceResourceId - The ID of the Amazon Web Services resource that was evaluated.

$sel:complianceType:Evaluation', evaluation_complianceType - Indicates whether the Amazon Web Services resource complies with the Config rule that it was evaluated against.

For the Evaluation data type, Config supports only the COMPLIANT, NON_COMPLIANT, and NOT_APPLICABLE values. Config does not support the INSUFFICIENT_DATA value for this data type.

Similarly, Config does not accept INSUFFICIENT_DATA as the value for ComplianceType from a PutEvaluations request. For example, an Lambda function for a custom Config rule cannot pass an INSUFFICIENT_DATA value to Config.

$sel:orderingTimestamp:Evaluation', evaluation_orderingTimestamp - The time of the event in Config that triggered the evaluation. For event-based evaluations, the time indicates when Config created the configuration item that triggered the evaluation. For periodic evaluations, the time indicates when Config triggered the evaluation at the frequency that you specified (for example, every 24 hours).

evaluation_annotation :: Lens' Evaluation (Maybe Text)

Supplementary information about how the evaluation determined the compliance.

evaluation_complianceResourceType :: Lens' Evaluation Text

The type of Amazon Web Services resource that was evaluated.

evaluation_complianceResourceId :: Lens' Evaluation Text

The ID of the Amazon Web Services resource that was evaluated.

evaluation_complianceType :: Lens' Evaluation ComplianceType

Indicates whether the Amazon Web Services resource complies with the Config rule that it was evaluated against.

For the Evaluation data type, Config supports only the COMPLIANT, NON_COMPLIANT, and NOT_APPLICABLE values. Config does not support the INSUFFICIENT_DATA value for this data type.

Similarly, Config does not accept INSUFFICIENT_DATA as the value for ComplianceType from a PutEvaluations request. For example, an Lambda function for a custom Config rule cannot pass an INSUFFICIENT_DATA value to Config.

evaluation_orderingTimestamp :: Lens' Evaluation UTCTime

The time of the event in Config that triggered the evaluation. For event-based evaluations, the time indicates when Config created the configuration item that triggered the evaluation. For periodic evaluations, the time indicates when Config triggered the evaluation at the frequency that you specified (for example, every 24 hours).