Amazonka.Config.Types.RemediationException

Description

Documentation

data RemediationException

An object that represents the details about the remediation exception. The details include the rule name, an explanation of an exception, the time when the exception will be deleted, the resource ID, and resource type.

See: newRemediationException smart constructor.

Constructors

RemediationException'

Fields

message :: Maybe Text

An explanation of an remediation exception.

expirationTime :: Maybe POSIX

The time when the remediation exception will be deleted.

configRuleName :: Text

The name of the Config rule.

resourceType :: Text

The type of a resource.

resourceId :: Text

The ID of the resource (for example., sg-xxxxxx).

newRemediationException

Arguments

:: Text
$sel:configRuleName:RemediationException'

-> Text
$sel:resourceType:RemediationException'

-> Text
$sel:resourceId:RemediationException'

-> RemediationException

Create a value of RemediationException with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:message:RemediationException', remediationException_message - An explanation of an remediation exception.

$sel:expirationTime:RemediationException', remediationException_expirationTime - The time when the remediation exception will be deleted.

$sel:configRuleName:RemediationException', remediationException_configRuleName - The name of the Config rule.

$sel:resourceType:RemediationException', remediationException_resourceType - The type of a resource.

$sel:resourceId:RemediationException', remediationException_resourceId - The ID of the resource (for example., sg-xxxxxx).

remediationException_message :: Lens' RemediationException (Maybe Text)

An explanation of an remediation exception.

remediationException_expirationTime :: Lens' RemediationException (Maybe UTCTime)

The time when the remediation exception will be deleted.

remediationException_configRuleName :: Lens' RemediationException Text

The name of the Config rule.

remediationException_resourceType :: Lens' RemediationException Text

The type of a resource.

remediationException_resourceId :: Lens' RemediationException Text

The ID of the resource (for example., sg-xxxxxx).