Amazonka.Config.Types.ResourceCount

Description

Documentation

data ResourceCount

An object that contains the resource type and the number of resources.

See: newResourceCount smart constructor.

Constructors

ResourceCount'

Fields

resourceType :: Maybe ResourceType

The resource type (for example, "AWS::EC2::Instance").

count :: Maybe Integer

The number of resources.

newResourceCount :: ResourceCount

Create a value of ResourceCount with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:resourceType:ResourceCount', resourceCount_resourceType - The resource type (for example, "AWS::EC2::Instance").

$sel:count:ResourceCount', resourceCount_count - The number of resources.

resourceCount_resourceType :: Lens' ResourceCount (Maybe ResourceType)

The resource type (for example, "AWS::EC2::Instance").

resourceCount_count :: Lens' ResourceCount (Maybe Integer)

The number of resources.