Amazonka.Connect.CreateInstance

Description

This API is in preview release for Amazon Connect and is subject to change.

Initiates an Amazon Connect instance with all the supported channels enabled. It does not attach any storage, such as Amazon Simple Storage Service (Amazon S3) or Amazon Kinesis. It also does not allow for any configurations on features, such as Contact Lens for Amazon Connect.

Amazon Connect enforces a limit on the total number of instances that you can create or delete in 30 days. If you exceed this limit, you will get an error message indicating there has been an excessive number of attempts at creating or deleting instances. You must wait 30 days before you can restart creating and deleting instances in your account.

Creating a Request

data CreateInstance

See: newCreateInstance smart constructor.

Constructors

CreateInstance'

Fields

directoryId :: Maybe Text

The identifier for the directory.

clientToken :: Maybe Text

The idempotency token.

instanceAlias :: Maybe (Sensitive Text)

The name for your instance.

identityManagementType :: DirectoryType

The type of identity management for your Amazon Connect users.

inboundCallsEnabled :: Bool

Your contact center handles incoming contacts.

outboundCallsEnabled :: Bool

Your contact center allows outbound calls.

newCreateInstance

Arguments

:: DirectoryType
$sel:identityManagementType:CreateInstance'

-> Bool
$sel:inboundCallsEnabled:CreateInstance'

-> Bool
$sel:outboundCallsEnabled:CreateInstance'

-> CreateInstance

Create a value of CreateInstance with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:directoryId:CreateInstance', createInstance_directoryId - The identifier for the directory.

$sel:clientToken:CreateInstance', createInstance_clientToken - The idempotency token.

$sel:instanceAlias:CreateInstance', createInstance_instanceAlias - The name for your instance.

$sel:identityManagementType:CreateInstance', createInstance_identityManagementType - The type of identity management for your Amazon Connect users.

$sel:inboundCallsEnabled:CreateInstance', createInstance_inboundCallsEnabled - Your contact center handles incoming contacts.

$sel:outboundCallsEnabled:CreateInstance', createInstance_outboundCallsEnabled - Your contact center allows outbound calls.

Request Lenses

createInstance_directoryId :: Lens' CreateInstance (Maybe Text)

The identifier for the directory.

createInstance_clientToken :: Lens' CreateInstance (Maybe Text)

The idempotency token.

createInstance_instanceAlias :: Lens' CreateInstance (Maybe Text)

The name for your instance.

createInstance_identityManagementType :: Lens' CreateInstance DirectoryType

The type of identity management for your Amazon Connect users.

createInstance_inboundCallsEnabled :: Lens' CreateInstance Bool

Your contact center handles incoming contacts.

createInstance_outboundCallsEnabled :: Lens' CreateInstance Bool

Your contact center allows outbound calls.

Destructuring the Response

data CreateInstanceResponse

See: newCreateInstanceResponse smart constructor.

Constructors

CreateInstanceResponse'

Fields

arn :: Maybe Text

The Amazon Resource Name (ARN) of the instance.

id :: Maybe Text

The identifier for the instance.

httpStatus :: Int

The response's http status code.

newCreateInstanceResponse

Arguments

:: Int
$sel:httpStatus:CreateInstanceResponse'

-> CreateInstanceResponse

Create a value of CreateInstanceResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:arn:CreateInstanceResponse', createInstanceResponse_arn - The Amazon Resource Name (ARN) of the instance.

$sel:id:CreateInstanceResponse', createInstanceResponse_id - The identifier for the instance.

$sel:httpStatus:CreateInstanceResponse', createInstanceResponse_httpStatus - The response's http status code.

Response Lenses

createInstanceResponse_arn :: Lens' CreateInstanceResponse (Maybe Text)

The Amazon Resource Name (ARN) of the instance.

createInstanceResponse_id :: Lens' CreateInstanceResponse (Maybe Text)

The identifier for the instance.

createInstanceResponse_httpStatus :: Lens' CreateInstanceResponse Int

The response's http status code.