Amazonka.Connect.CreateRoutingProfile

Description

Creates a new routing profile.

Creating a Request

data CreateRoutingProfile

See: newCreateRoutingProfile smart constructor.

Constructors

CreateRoutingProfile'

Fields

queueConfigs :: Maybe (NonEmpty RoutingProfileQueueConfig)

The inbound queues associated with the routing profile. If no queue is added, the agent can make only outbound calls.

tags :: Maybe (HashMap Text Text)

One or more tags.

instanceId :: Text

The identifier of the Amazon Connect instance. You can find the instanceId in the ARN of the instance.

name :: Text

The name of the routing profile. Must not be more than 127 characters.

description :: Text

Description of the routing profile. Must not be more than 250 characters.

defaultOutboundQueueId :: Text

The default outbound queue for the routing profile.

mediaConcurrencies :: [MediaConcurrency]

The channels that agents can handle in the Contact Control Panel (CCP) for this routing profile.

newCreateRoutingProfile

Arguments

:: Text
$sel:instanceId:CreateRoutingProfile'

-> Text
$sel:name:CreateRoutingProfile'

-> Text
$sel:description:CreateRoutingProfile'

-> Text
$sel:defaultOutboundQueueId:CreateRoutingProfile'

-> CreateRoutingProfile

Create a value of CreateRoutingProfile with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:queueConfigs:CreateRoutingProfile', createRoutingProfile_queueConfigs - The inbound queues associated with the routing profile. If no queue is added, the agent can make only outbound calls.

$sel:tags:CreateRoutingProfile', createRoutingProfile_tags - One or more tags.

$sel:instanceId:CreateRoutingProfile', createRoutingProfile_instanceId - The identifier of the Amazon Connect instance. You can find the instanceId in the ARN of the instance.

$sel:name:CreateRoutingProfile', createRoutingProfile_name - The name of the routing profile. Must not be more than 127 characters.

$sel:description:CreateRoutingProfile', createRoutingProfile_description - Description of the routing profile. Must not be more than 250 characters.

$sel:defaultOutboundQueueId:CreateRoutingProfile', createRoutingProfile_defaultOutboundQueueId - The default outbound queue for the routing profile.

$sel:mediaConcurrencies:CreateRoutingProfile', createRoutingProfile_mediaConcurrencies - The channels that agents can handle in the Contact Control Panel (CCP) for this routing profile.

Request Lenses

createRoutingProfile_queueConfigs :: Lens' CreateRoutingProfile (Maybe (NonEmpty RoutingProfileQueueConfig))

The inbound queues associated with the routing profile. If no queue is added, the agent can make only outbound calls.

createRoutingProfile_tags :: Lens' CreateRoutingProfile (Maybe (HashMap Text Text))

One or more tags.

createRoutingProfile_instanceId :: Lens' CreateRoutingProfile Text

The identifier of the Amazon Connect instance. You can find the instanceId in the ARN of the instance.

createRoutingProfile_name :: Lens' CreateRoutingProfile Text

The name of the routing profile. Must not be more than 127 characters.

createRoutingProfile_description :: Lens' CreateRoutingProfile Text

Description of the routing profile. Must not be more than 250 characters.

createRoutingProfile_defaultOutboundQueueId :: Lens' CreateRoutingProfile Text

The default outbound queue for the routing profile.

createRoutingProfile_mediaConcurrencies :: Lens' CreateRoutingProfile [MediaConcurrency]

The channels that agents can handle in the Contact Control Panel (CCP) for this routing profile.

Destructuring the Response

data CreateRoutingProfileResponse

See: newCreateRoutingProfileResponse smart constructor.

Constructors

CreateRoutingProfileResponse'

Fields

routingProfileArn :: Maybe Text

The Amazon Resource Name (ARN) of the routing profile.

routingProfileId :: Maybe Text

The identifier of the routing profile.

httpStatus :: Int

The response's http status code.

newCreateRoutingProfileResponse

Arguments

:: Int
$sel:httpStatus:CreateRoutingProfileResponse'

-> CreateRoutingProfileResponse

Create a value of CreateRoutingProfileResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:routingProfileArn:CreateRoutingProfileResponse', createRoutingProfileResponse_routingProfileArn - The Amazon Resource Name (ARN) of the routing profile.

$sel:routingProfileId:CreateRoutingProfileResponse', createRoutingProfileResponse_routingProfileId - The identifier of the routing profile.

$sel:httpStatus:CreateRoutingProfileResponse', createRoutingProfileResponse_httpStatus - The response's http status code.

Response Lenses

createRoutingProfileResponse_routingProfileArn :: Lens' CreateRoutingProfileResponse (Maybe Text)

The Amazon Resource Name (ARN) of the routing profile.

createRoutingProfileResponse_routingProfileId :: Lens' CreateRoutingProfileResponse (Maybe Text)

The identifier of the routing profile.

createRoutingProfileResponse_httpStatus :: Lens' CreateRoutingProfileResponse Int

The response's http status code.