Amazonka.Connect.DescribeContactFlow

Description

Describes the specified contact flow.

You can also create and update contact flows using the Amazon Connect Flow language.

Creating a Request

data DescribeContactFlow

See: newDescribeContactFlow smart constructor.

Constructors

DescribeContactFlow'

Fields

instanceId :: Text

The identifier of the Amazon Connect instance.

contactFlowId :: Text

The identifier of the contact flow.

newDescribeContactFlow

Arguments

:: Text
$sel:instanceId:DescribeContactFlow'

-> Text
$sel:contactFlowId:DescribeContactFlow'

-> DescribeContactFlow

Create a value of DescribeContactFlow with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:instanceId:DescribeContactFlow', describeContactFlow_instanceId - The identifier of the Amazon Connect instance.

$sel:contactFlowId:DescribeContactFlow', describeContactFlow_contactFlowId - The identifier of the contact flow.

Request Lenses

describeContactFlow_instanceId :: Lens' DescribeContactFlow Text

The identifier of the Amazon Connect instance.

describeContactFlow_contactFlowId :: Lens' DescribeContactFlow Text

The identifier of the contact flow.

Destructuring the Response

data DescribeContactFlowResponse

See: newDescribeContactFlowResponse smart constructor.

Constructors

DescribeContactFlowResponse'

Fields

contactFlow :: Maybe ContactFlow

Information about the contact flow.

httpStatus :: Int

The response's http status code.

newDescribeContactFlowResponse

Arguments

:: Int
$sel:httpStatus:DescribeContactFlowResponse'

-> DescribeContactFlowResponse

Create a value of DescribeContactFlowResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:contactFlow:DescribeContactFlowResponse', describeContactFlowResponse_contactFlow - Information about the contact flow.

$sel:httpStatus:DescribeContactFlowResponse', describeContactFlowResponse_httpStatus - The response's http status code.

Response Lenses

describeContactFlowResponse_contactFlow :: Lens' DescribeContactFlowResponse (Maybe ContactFlow)

Information about the contact flow.

describeContactFlowResponse_httpStatus :: Lens' DescribeContactFlowResponse Int

The response's http status code.