Amazonka.Connect.DescribeQueue

Description

This API is in preview release for Amazon Connect and is subject to change.

Describes the specified queue.

Creating a Request

data DescribeQueue

See: newDescribeQueue smart constructor.

Constructors

DescribeQueue'

Fields

instanceId :: Text

The identifier of the Amazon Connect instance. You can find the instanceId in the ARN of the instance.

queueId :: Text

The identifier for the queue.

newDescribeQueue

Arguments

:: Text
$sel:instanceId:DescribeQueue'

-> Text
$sel:queueId:DescribeQueue'

-> DescribeQueue

Create a value of DescribeQueue with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:instanceId:DescribeQueue', describeQueue_instanceId - The identifier of the Amazon Connect instance. You can find the instanceId in the ARN of the instance.

$sel:queueId:DescribeQueue', describeQueue_queueId - The identifier for the queue.

Request Lenses

describeQueue_instanceId :: Lens' DescribeQueue Text

The identifier of the Amazon Connect instance. You can find the instanceId in the ARN of the instance.

describeQueue_queueId :: Lens' DescribeQueue Text

The identifier for the queue.

Destructuring the Response

data DescribeQueueResponse

See: newDescribeQueueResponse smart constructor.

Constructors

DescribeQueueResponse'

Fields

queue :: Maybe Queue

The name of the queue.

httpStatus :: Int

The response's http status code.

newDescribeQueueResponse

Arguments

:: Int
$sel:httpStatus:DescribeQueueResponse'

-> DescribeQueueResponse

Create a value of DescribeQueueResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:queue:DescribeQueueResponse', describeQueueResponse_queue - The name of the queue.

$sel:httpStatus:DescribeQueueResponse', describeQueueResponse_httpStatus - The response's http status code.

Response Lenses

describeQueueResponse_queue :: Lens' DescribeQueueResponse (Maybe Queue)

The name of the queue.

describeQueueResponse_httpStatus :: Lens' DescribeQueueResponse Int

The response's http status code.