Amazonka.Connect.DescribeQuickConnect

Description

Describes the quick connect.

Creating a Request

data DescribeQuickConnect

See: newDescribeQuickConnect smart constructor.

Constructors

DescribeQuickConnect'

Fields

instanceId :: Text

The identifier of the Amazon Connect instance. You can find the instanceId in the ARN of the instance.

quickConnectId :: Text

The identifier for the quick connect.

newDescribeQuickConnect

Arguments

:: Text
$sel:instanceId:DescribeQuickConnect'

-> Text
$sel:quickConnectId:DescribeQuickConnect'

-> DescribeQuickConnect

Create a value of DescribeQuickConnect with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:instanceId:DescribeQuickConnect', describeQuickConnect_instanceId - The identifier of the Amazon Connect instance. You can find the instanceId in the ARN of the instance.

$sel:quickConnectId:DescribeQuickConnect', describeQuickConnect_quickConnectId - The identifier for the quick connect.

Request Lenses

describeQuickConnect_instanceId :: Lens' DescribeQuickConnect Text

The identifier of the Amazon Connect instance. You can find the instanceId in the ARN of the instance.

describeQuickConnect_quickConnectId :: Lens' DescribeQuickConnect Text

The identifier for the quick connect.

Destructuring the Response

data DescribeQuickConnectResponse

See: newDescribeQuickConnectResponse smart constructor.

Constructors

DescribeQuickConnectResponse'

Fields

quickConnect :: Maybe QuickConnect

Information about the quick connect.

httpStatus :: Int

The response's http status code.

newDescribeQuickConnectResponse

Arguments

:: Int
$sel:httpStatus:DescribeQuickConnectResponse'

-> DescribeQuickConnectResponse

Create a value of DescribeQuickConnectResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:quickConnect:DescribeQuickConnectResponse', describeQuickConnectResponse_quickConnect - Information about the quick connect.

$sel:httpStatus:DescribeQuickConnectResponse', describeQuickConnectResponse_httpStatus - The response's http status code.

Response Lenses

describeQuickConnectResponse_quickConnect :: Lens' DescribeQuickConnectResponse (Maybe QuickConnect)

Information about the quick connect.

describeQuickConnectResponse_httpStatus :: Lens' DescribeQuickConnectResponse Int

The response's http status code.