Amazonka.Connect.DisassociateBot

Description

This API is in preview release for Amazon Connect and is subject to change.

Revokes authorization from the specified instance to access the specified Amazon Lex or Amazon Lex V2 bot.

Creating a Request

data DisassociateBot

See: newDisassociateBot smart constructor.

Constructors

DisassociateBot'

Fields

lexBot :: Maybe LexBot

lexV2Bot :: Maybe LexV2Bot

The Amazon Lex V2 bot to disassociate from the instance.

instanceId :: Text

The identifier of the Amazon Connect instance. You can find the instanceId in the ARN of the instance.

newDisassociateBot

Arguments

:: Text
$sel:instanceId:DisassociateBot'

-> DisassociateBot

Create a value of DisassociateBot with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:lexBot:DisassociateBot', disassociateBot_lexBot - Undocumented member.

$sel:lexV2Bot:DisassociateBot', disassociateBot_lexV2Bot - The Amazon Lex V2 bot to disassociate from the instance.

$sel:instanceId:DisassociateBot', disassociateBot_instanceId - The identifier of the Amazon Connect instance. You can find the instanceId in the ARN of the instance.

Request Lenses

disassociateBot_lexBot :: Lens' DisassociateBot (Maybe LexBot)

Undocumented member.

disassociateBot_lexV2Bot :: Lens' DisassociateBot (Maybe LexV2Bot)

The Amazon Lex V2 bot to disassociate from the instance.

disassociateBot_instanceId :: Lens' DisassociateBot Text

The identifier of the Amazon Connect instance. You can find the instanceId in the ARN of the instance.

Destructuring the Response

data DisassociateBotResponse

See: newDisassociateBotResponse smart constructor.

Constructors

DisassociateBotResponse'

newDisassociateBotResponse :: DisassociateBotResponse

Create a value of DisassociateBotResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.