Amazonka.Connect.DisassociateQueueQuickConnects

Description

This API is in preview release for Amazon Connect and is subject to change.

Disassociates a set of quick connects from a queue.

Creating a Request

data DisassociateQueueQuickConnects

See: newDisassociateQueueQuickConnects smart constructor.

Constructors

DisassociateQueueQuickConnects'

Fields

instanceId :: Text

The identifier of the Amazon Connect instance. You can find the instanceId in the ARN of the instance.

queueId :: Text

The identifier for the queue.

quickConnectIds :: NonEmpty Text

The quick connects to disassociate from the queue.

newDisassociateQueueQuickConnects

Arguments

:: Text
$sel:instanceId:DisassociateQueueQuickConnects'

-> Text
$sel:queueId:DisassociateQueueQuickConnects'

-> NonEmpty Text
$sel:quickConnectIds:DisassociateQueueQuickConnects'

-> DisassociateQueueQuickConnects

Create a value of DisassociateQueueQuickConnects with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:instanceId:DisassociateQueueQuickConnects', disassociateQueueQuickConnects_instanceId - The identifier of the Amazon Connect instance. You can find the instanceId in the ARN of the instance.

$sel:queueId:DisassociateQueueQuickConnects', disassociateQueueQuickConnects_queueId - The identifier for the queue.

$sel:quickConnectIds:DisassociateQueueQuickConnects', disassociateQueueQuickConnects_quickConnectIds - The quick connects to disassociate from the queue.

Request Lenses

disassociateQueueQuickConnects_instanceId :: Lens' DisassociateQueueQuickConnects Text

The identifier of the Amazon Connect instance. You can find the instanceId in the ARN of the instance.

disassociateQueueQuickConnects_queueId :: Lens' DisassociateQueueQuickConnects Text

The identifier for the queue.

disassociateQueueQuickConnects_quickConnectIds :: Lens' DisassociateQueueQuickConnects (NonEmpty Text)

The quick connects to disassociate from the queue.

Destructuring the Response

data DisassociateQueueQuickConnectsResponse

See: newDisassociateQueueQuickConnectsResponse smart constructor.

Constructors

DisassociateQueueQuickConnectsResponse'

newDisassociateQueueQuickConnectsResponse :: DisassociateQueueQuickConnectsResponse

Create a value of DisassociateQueueQuickConnectsResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.