Amazonka.Connect.GetCurrentMetricData

Description

Gets the real-time metric data from the specified Amazon Connect instance.

For a description of each metric, see Real-time Metrics Definitions in the Amazon Connect Administrator Guide.

Creating a Request

data GetCurrentMetricData

See: newGetCurrentMetricData smart constructor.

Constructors

GetCurrentMetricData'

Fields

nextToken :: Maybe Text

The token for the next set of results. Use the value returned in the previous response in the next request to retrieve the next set of results.

The token expires after 5 minutes from the time it is created. Subsequent requests that use the token must use the same request parameters as the request that generated the token.

groupings :: Maybe [Grouping]

The grouping applied to the metrics returned. For example, when grouped by QUEUE, the metrics returned apply to each queue rather than aggregated for all queues. If you group by CHANNEL, you should include a Channels filter. VOICE, CHAT, and TASK channels are supported.

If no Grouping is included in the request, a summary of metrics is returned.

maxResults :: Maybe Natural

The maximum number of results to return per page.

instanceId :: Text

The identifier of the Amazon Connect instance. You can find the instanceId in the ARN of the instance.

filters :: Filters

The queues, up to 100, or channels, to use to filter the metrics returned. Metric data is retrieved only for the resources associated with the queues or channels included in the filter. You can include both queue IDs and queue ARNs in the same request. VOICE, CHAT, and TASK channels are supported.

currentMetrics :: [CurrentMetric]

The metrics to retrieve. Specify the name and unit for each metric. The following metrics are available. For a description of all the metrics, see Real-time Metrics Definitions in the Amazon Connect Administrator Guide.

AGENTS_AFTER_CONTACT_WORK

Unit: COUNT
Name in real-time metrics report: ACW

AGENTS_AVAILABLE

Unit: COUNT
Name in real-time metrics report: Available

AGENTS_ERROR

Unit: COUNT
Name in real-time metrics report: Error

AGENTS_NON_PRODUCTIVE

Unit: COUNT
Name in real-time metrics report: NPT (Non-Productive Time)

AGENTS_ON_CALL

Unit: COUNT
Name in real-time metrics report: On contact

AGENTS_ON_CONTACT

Unit: COUNT
Name in real-time metrics report: On contact

AGENTS_ONLINE

Unit: COUNT
Name in real-time metrics report: Online

AGENTS_STAFFED

Unit: COUNT
Name in real-time metrics report: Staffed

CONTACTS_IN_QUEUE

Unit: COUNT
Name in real-time metrics report: In queue

CONTACTS_SCHEDULED

Unit: COUNT
Name in real-time metrics report: Scheduled

OLDEST_CONTACT_AGE

Unit: SECONDS
When you use groupings, Unit says SECONDS and the Value is returned in SECONDS.

When you do not use groupings, Unit says SECONDS but the Value is returned in MILLISECONDS. For example, if you get a response like this:

{ "Metric": { "Name": "OLDEST_CONTACT_AGE", "Unit": "SECONDS" }, "Value": 24113.0 }

The actual OLDEST_CONTACT_AGE is 24 seconds.

Name in real-time metrics report: Oldest

SLOTS_ACTIVE

Unit: COUNT
Name in real-time metrics report: Active

SLOTS_AVAILABLE

Unit: COUNT
Name in real-time metrics report: Availability

newGetCurrentMetricData

Arguments

:: Text
$sel:instanceId:GetCurrentMetricData'

-> Filters
$sel:filters:GetCurrentMetricData'

-> GetCurrentMetricData

Create a value of GetCurrentMetricData with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:nextToken:GetCurrentMetricData', getCurrentMetricData_nextToken - The token for the next set of results. Use the value returned in the previous response in the next request to retrieve the next set of results.

The token expires after 5 minutes from the time it is created. Subsequent requests that use the token must use the same request parameters as the request that generated the token.

$sel:groupings:GetCurrentMetricData', getCurrentMetricData_groupings - The grouping applied to the metrics returned. For example, when grouped by QUEUE, the metrics returned apply to each queue rather than aggregated for all queues. If you group by CHANNEL, you should include a Channels filter. VOICE, CHAT, and TASK channels are supported.

If no Grouping is included in the request, a summary of metrics is returned.

$sel:maxResults:GetCurrentMetricData', getCurrentMetricData_maxResults - The maximum number of results to return per page.

$sel:instanceId:GetCurrentMetricData', getCurrentMetricData_instanceId - The identifier of the Amazon Connect instance. You can find the instanceId in the ARN of the instance.

$sel:filters:GetCurrentMetricData', getCurrentMetricData_filters - The queues, up to 100, or channels, to use to filter the metrics returned. Metric data is retrieved only for the resources associated with the queues or channels included in the filter. You can include both queue IDs and queue ARNs in the same request. VOICE, CHAT, and TASK channels are supported.

$sel:currentMetrics:GetCurrentMetricData', getCurrentMetricData_currentMetrics - The metrics to retrieve. Specify the name and unit for each metric. The following metrics are available. For a description of all the metrics, see Real-time Metrics Definitions in the Amazon Connect Administrator Guide.

AGENTS_AFTER_CONTACT_WORK

Unit: COUNT
Name in real-time metrics report: ACW

AGENTS_AVAILABLE

Unit: COUNT
Name in real-time metrics report: Available

AGENTS_ERROR

Unit: COUNT
Name in real-time metrics report: Error

AGENTS_NON_PRODUCTIVE

Unit: COUNT
Name in real-time metrics report: NPT (Non-Productive Time)

AGENTS_ON_CALL

Unit: COUNT
Name in real-time metrics report: On contact

AGENTS_ON_CONTACT

Unit: COUNT
Name in real-time metrics report: On contact

AGENTS_ONLINE

Unit: COUNT
Name in real-time metrics report: Online

AGENTS_STAFFED

Unit: COUNT
Name in real-time metrics report: Staffed

CONTACTS_IN_QUEUE

Unit: COUNT
Name in real-time metrics report: In queue

CONTACTS_SCHEDULED

Unit: COUNT
Name in real-time metrics report: Scheduled

OLDEST_CONTACT_AGE

Unit: SECONDS
When you use groupings, Unit says SECONDS and the Value is returned in SECONDS.

When you do not use groupings, Unit says SECONDS but the Value is returned in MILLISECONDS. For example, if you get a response like this:

{ "Metric": { "Name": "OLDEST_CONTACT_AGE", "Unit": "SECONDS" }, "Value": 24113.0 }

The actual OLDEST_CONTACT_AGE is 24 seconds.

Name in real-time metrics report: Oldest

SLOTS_ACTIVE

Unit: COUNT
Name in real-time metrics report: Active

SLOTS_AVAILABLE

Unit: COUNT
Name in real-time metrics report: Availability

Request Lenses

getCurrentMetricData_nextToken :: Lens' GetCurrentMetricData (Maybe Text)

The token for the next set of results. Use the value returned in the previous response in the next request to retrieve the next set of results.

The token expires after 5 minutes from the time it is created. Subsequent requests that use the token must use the same request parameters as the request that generated the token.

getCurrentMetricData_groupings :: Lens' GetCurrentMetricData (Maybe [Grouping])

The grouping applied to the metrics returned. For example, when grouped by QUEUE, the metrics returned apply to each queue rather than aggregated for all queues. If you group by CHANNEL, you should include a Channels filter. VOICE, CHAT, and TASK channels are supported.

If no Grouping is included in the request, a summary of metrics is returned.

getCurrentMetricData_maxResults :: Lens' GetCurrentMetricData (Maybe Natural)

The maximum number of results to return per page.

getCurrentMetricData_instanceId :: Lens' GetCurrentMetricData Text

The identifier of the Amazon Connect instance. You can find the instanceId in the ARN of the instance.

getCurrentMetricData_filters :: Lens' GetCurrentMetricData Filters

The queues, up to 100, or channels, to use to filter the metrics returned. Metric data is retrieved only for the resources associated with the queues or channels included in the filter. You can include both queue IDs and queue ARNs in the same request. VOICE, CHAT, and TASK channels are supported.

getCurrentMetricData_currentMetrics :: Lens' GetCurrentMetricData [CurrentMetric]

The metrics to retrieve. Specify the name and unit for each metric. The following metrics are available. For a description of all the metrics, see Real-time Metrics Definitions in the Amazon Connect Administrator Guide.

AGENTS_AFTER_CONTACT_WORK

Unit: COUNT
Name in real-time metrics report: ACW

AGENTS_AVAILABLE

Unit: COUNT
Name in real-time metrics report: Available

AGENTS_ERROR

Unit: COUNT
Name in real-time metrics report: Error

AGENTS_NON_PRODUCTIVE

Unit: COUNT
Name in real-time metrics report: NPT (Non-Productive Time)

AGENTS_ON_CALL

Unit: COUNT
Name in real-time metrics report: On contact

AGENTS_ON_CONTACT

Unit: COUNT
Name in real-time metrics report: On contact

AGENTS_ONLINE

Unit: COUNT
Name in real-time metrics report: Online

AGENTS_STAFFED

Unit: COUNT
Name in real-time metrics report: Staffed

CONTACTS_IN_QUEUE

Unit: COUNT
Name in real-time metrics report: In queue

CONTACTS_SCHEDULED

Unit: COUNT
Name in real-time metrics report: Scheduled

OLDEST_CONTACT_AGE

Unit: SECONDS
When you use groupings, Unit says SECONDS and the Value is returned in SECONDS.

When you do not use groupings, Unit says SECONDS but the Value is returned in MILLISECONDS. For example, if you get a response like this:

{ "Metric": { "Name": "OLDEST_CONTACT_AGE", "Unit": "SECONDS" }, "Value": 24113.0 }

The actual OLDEST_CONTACT_AGE is 24 seconds.

Name in real-time metrics report: Oldest

SLOTS_ACTIVE

Unit: COUNT
Name in real-time metrics report: Active

SLOTS_AVAILABLE

Unit: COUNT
Name in real-time metrics report: Availability

Destructuring the Response

data GetCurrentMetricDataResponse

See: newGetCurrentMetricDataResponse smart constructor.

Constructors

GetCurrentMetricDataResponse'

Fields

metricResults :: Maybe [CurrentMetricResult]

Information about the real-time metrics.

dataSnapshotTime :: Maybe POSIX

The time at which the metrics were retrieved and cached for pagination.

nextToken :: Maybe Text

If there are additional results, this is the token for the next set of results.

The token expires after 5 minutes from the time it is created. Subsequent requests that use the token must use the same request parameters as the request that generated the token.

httpStatus :: Int

The response's http status code.

newGetCurrentMetricDataResponse

Arguments

:: Int
$sel:httpStatus:GetCurrentMetricDataResponse'

-> GetCurrentMetricDataResponse

Create a value of GetCurrentMetricDataResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:metricResults:GetCurrentMetricDataResponse', getCurrentMetricDataResponse_metricResults - Information about the real-time metrics.

$sel:dataSnapshotTime:GetCurrentMetricDataResponse', getCurrentMetricDataResponse_dataSnapshotTime - The time at which the metrics were retrieved and cached for pagination.

$sel:nextToken:GetCurrentMetricData', getCurrentMetricDataResponse_nextToken - If there are additional results, this is the token for the next set of results.

The token expires after 5 minutes from the time it is created. Subsequent requests that use the token must use the same request parameters as the request that generated the token.

$sel:httpStatus:GetCurrentMetricDataResponse', getCurrentMetricDataResponse_httpStatus - The response's http status code.

Response Lenses

getCurrentMetricDataResponse_metricResults :: Lens' GetCurrentMetricDataResponse (Maybe [CurrentMetricResult])

Information about the real-time metrics.

getCurrentMetricDataResponse_dataSnapshotTime :: Lens' GetCurrentMetricDataResponse (Maybe UTCTime)

The time at which the metrics were retrieved and cached for pagination.

getCurrentMetricDataResponse_nextToken :: Lens' GetCurrentMetricDataResponse (Maybe Text)

If there are additional results, this is the token for the next set of results.

The token expires after 5 minutes from the time it is created. Subsequent requests that use the token must use the same request parameters as the request that generated the token.

getCurrentMetricDataResponse_httpStatus :: Lens' GetCurrentMetricDataResponse Int

The response's http status code.