Amazonka.Connect.ListInstanceAttributes

Description

This API is in preview release for Amazon Connect and is subject to change.

Returns a paginated list of all attribute types for the given instance.

This operation returns paginated results.

Creating a Request

data ListInstanceAttributes

See: newListInstanceAttributes smart constructor.

Constructors

ListInstanceAttributes'

Fields

nextToken :: Maybe Text

The token for the next set of results. Use the value returned in the previous response in the next request to retrieve the next set of results.

maxResults :: Maybe Natural

The maximum number of results to return per page.

instanceId :: Text

The identifier of the Amazon Connect instance. You can find the instanceId in the ARN of the instance.

newListInstanceAttributes

Arguments

:: Text
$sel:instanceId:ListInstanceAttributes'

-> ListInstanceAttributes

Create a value of ListInstanceAttributes with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:nextToken:ListInstanceAttributes', listInstanceAttributes_nextToken - The token for the next set of results. Use the value returned in the previous response in the next request to retrieve the next set of results.

$sel:maxResults:ListInstanceAttributes', listInstanceAttributes_maxResults - The maximum number of results to return per page.

$sel:instanceId:ListInstanceAttributes', listInstanceAttributes_instanceId - The identifier of the Amazon Connect instance. You can find the instanceId in the ARN of the instance.

Request Lenses

listInstanceAttributes_nextToken :: Lens' ListInstanceAttributes (Maybe Text)

The token for the next set of results. Use the value returned in the previous response in the next request to retrieve the next set of results.

listInstanceAttributes_maxResults :: Lens' ListInstanceAttributes (Maybe Natural)

The maximum number of results to return per page.

listInstanceAttributes_instanceId :: Lens' ListInstanceAttributes Text

The identifier of the Amazon Connect instance. You can find the instanceId in the ARN of the instance.

Destructuring the Response

data ListInstanceAttributesResponse

See: newListInstanceAttributesResponse smart constructor.

Constructors

ListInstanceAttributesResponse'

Fields

nextToken :: Maybe Text

If there are additional results, this is the token for the next set of results.

attributes :: Maybe [Attribute]

The attribute types.

httpStatus :: Int

The response's http status code.

newListInstanceAttributesResponse

Arguments

:: Int
$sel:httpStatus:ListInstanceAttributesResponse'

-> ListInstanceAttributesResponse

Create a value of ListInstanceAttributesResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:nextToken:ListInstanceAttributes', listInstanceAttributesResponse_nextToken - If there are additional results, this is the token for the next set of results.

$sel:attributes:ListInstanceAttributesResponse', listInstanceAttributesResponse_attributes - The attribute types.

$sel:httpStatus:ListInstanceAttributesResponse', listInstanceAttributesResponse_httpStatus - The response's http status code.

Response Lenses

listInstanceAttributesResponse_nextToken :: Lens' ListInstanceAttributesResponse (Maybe Text)

If there are additional results, this is the token for the next set of results.

listInstanceAttributesResponse_attributes :: Lens' ListInstanceAttributesResponse (Maybe [Attribute])

The attribute types.

listInstanceAttributesResponse_httpStatus :: Lens' ListInstanceAttributesResponse Int

The response's http status code.