Amazonka.Connect.ListPhoneNumbers

Description

Provides information about the phone numbers for the specified Amazon Connect instance.

For more information about phone numbers, see Set Up Phone Numbers for Your Contact Center in the Amazon Connect Administrator Guide.

This operation returns paginated results.

Creating a Request

data ListPhoneNumbers

See: newListPhoneNumbers smart constructor.

Constructors

ListPhoneNumbers'

Fields

phoneNumberTypes :: Maybe [PhoneNumberType]

The type of phone number.

phoneNumberCountryCodes :: Maybe [PhoneNumberCountryCode]

The ISO country code.

nextToken :: Maybe Text

The token for the next set of results. Use the value returned in the previous response in the next request to retrieve the next set of results.

maxResults :: Maybe Natural

The maximum number of results to return per page.

instanceId :: Text

The identifier of the Amazon Connect instance. You can find the instanceId in the ARN of the instance.

newListPhoneNumbers

Arguments

:: Text
$sel:instanceId:ListPhoneNumbers'

-> ListPhoneNumbers

Create a value of ListPhoneNumbers with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:phoneNumberTypes:ListPhoneNumbers', listPhoneNumbers_phoneNumberTypes - The type of phone number.

$sel:phoneNumberCountryCodes:ListPhoneNumbers', listPhoneNumbers_phoneNumberCountryCodes - The ISO country code.

$sel:nextToken:ListPhoneNumbers', listPhoneNumbers_nextToken - The token for the next set of results. Use the value returned in the previous response in the next request to retrieve the next set of results.

$sel:maxResults:ListPhoneNumbers', listPhoneNumbers_maxResults - The maximum number of results to return per page.

$sel:instanceId:ListPhoneNumbers', listPhoneNumbers_instanceId - The identifier of the Amazon Connect instance. You can find the instanceId in the ARN of the instance.

Request Lenses

listPhoneNumbers_phoneNumberTypes :: Lens' ListPhoneNumbers (Maybe [PhoneNumberType])

The type of phone number.

listPhoneNumbers_phoneNumberCountryCodes :: Lens' ListPhoneNumbers (Maybe [PhoneNumberCountryCode])

The ISO country code.

listPhoneNumbers_nextToken :: Lens' ListPhoneNumbers (Maybe Text)

The token for the next set of results. Use the value returned in the previous response in the next request to retrieve the next set of results.

listPhoneNumbers_maxResults :: Lens' ListPhoneNumbers (Maybe Natural)

The maximum number of results to return per page.

listPhoneNumbers_instanceId :: Lens' ListPhoneNumbers Text

The identifier of the Amazon Connect instance. You can find the instanceId in the ARN of the instance.

Destructuring the Response

data ListPhoneNumbersResponse

See: newListPhoneNumbersResponse smart constructor.

Constructors

ListPhoneNumbersResponse'

Fields

phoneNumberSummaryList :: Maybe [PhoneNumberSummary]

Information about the phone numbers.

nextToken :: Maybe Text

If there are additional results, this is the token for the next set of results.

httpStatus :: Int

The response's http status code.

newListPhoneNumbersResponse

Arguments

:: Int
$sel:httpStatus:ListPhoneNumbersResponse'

-> ListPhoneNumbersResponse

Create a value of ListPhoneNumbersResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:phoneNumberSummaryList:ListPhoneNumbersResponse', listPhoneNumbersResponse_phoneNumberSummaryList - Information about the phone numbers.

$sel:nextToken:ListPhoneNumbers', listPhoneNumbersResponse_nextToken - If there are additional results, this is the token for the next set of results.

$sel:httpStatus:ListPhoneNumbersResponse', listPhoneNumbersResponse_httpStatus - The response's http status code.

Response Lenses

listPhoneNumbersResponse_phoneNumberSummaryList :: Lens' ListPhoneNumbersResponse (Maybe [PhoneNumberSummary])

Information about the phone numbers.

listPhoneNumbersResponse_nextToken :: Lens' ListPhoneNumbersResponse (Maybe Text)

If there are additional results, this is the token for the next set of results.

listPhoneNumbersResponse_httpStatus :: Lens' ListPhoneNumbersResponse Int

The response's http status code.