Amazonka.Connect.ListSecurityProfiles

Description

Provides summary information about the security profiles for the specified Amazon Connect instance.

For more information about security profiles, see Security Profiles in the Amazon Connect Administrator Guide.

This operation returns paginated results.

Creating a Request

data ListSecurityProfiles

See: newListSecurityProfiles smart constructor.

Constructors

ListSecurityProfiles'

Fields

nextToken :: Maybe Text

The token for the next set of results. Use the value returned in the previous response in the next request to retrieve the next set of results.

maxResults :: Maybe Natural

The maximum number of results to return per page.

instanceId :: Text

The identifier of the Amazon Connect instance. You can find the instanceId in the ARN of the instance.

newListSecurityProfiles

Arguments

:: Text
$sel:instanceId:ListSecurityProfiles'

-> ListSecurityProfiles

Create a value of ListSecurityProfiles with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:nextToken:ListSecurityProfiles', listSecurityProfiles_nextToken - The token for the next set of results. Use the value returned in the previous response in the next request to retrieve the next set of results.

$sel:maxResults:ListSecurityProfiles', listSecurityProfiles_maxResults - The maximum number of results to return per page.

$sel:instanceId:ListSecurityProfiles', listSecurityProfiles_instanceId - The identifier of the Amazon Connect instance. You can find the instanceId in the ARN of the instance.

Request Lenses

listSecurityProfiles_nextToken :: Lens' ListSecurityProfiles (Maybe Text)

The token for the next set of results. Use the value returned in the previous response in the next request to retrieve the next set of results.

listSecurityProfiles_maxResults :: Lens' ListSecurityProfiles (Maybe Natural)

The maximum number of results to return per page.

listSecurityProfiles_instanceId :: Lens' ListSecurityProfiles Text

The identifier of the Amazon Connect instance. You can find the instanceId in the ARN of the instance.

Destructuring the Response

data ListSecurityProfilesResponse

See: newListSecurityProfilesResponse smart constructor.

Constructors

ListSecurityProfilesResponse'

Fields

nextToken :: Maybe Text

If there are additional results, this is the token for the next set of results.

securityProfileSummaryList :: Maybe [SecurityProfileSummary]

Information about the security profiles.

httpStatus :: Int

The response's http status code.

newListSecurityProfilesResponse

Arguments

:: Int
$sel:httpStatus:ListSecurityProfilesResponse'

-> ListSecurityProfilesResponse

Create a value of ListSecurityProfilesResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:nextToken:ListSecurityProfiles', listSecurityProfilesResponse_nextToken - If there are additional results, this is the token for the next set of results.

$sel:securityProfileSummaryList:ListSecurityProfilesResponse', listSecurityProfilesResponse_securityProfileSummaryList - Information about the security profiles.

$sel:httpStatus:ListSecurityProfilesResponse', listSecurityProfilesResponse_httpStatus - The response's http status code.

Response Lenses

listSecurityProfilesResponse_nextToken :: Lens' ListSecurityProfilesResponse (Maybe Text)

If there are additional results, this is the token for the next set of results.

listSecurityProfilesResponse_securityProfileSummaryList :: Lens' ListSecurityProfilesResponse (Maybe [SecurityProfileSummary])

Information about the security profiles.

listSecurityProfilesResponse_httpStatus :: Lens' ListSecurityProfilesResponse Int

The response's http status code.