Amazonka.Connect.StartTaskContact

Description

Initiates a contact flow to start a new task.

Creating a Request

data StartTaskContact

See: newStartTaskContact smart constructor.

Constructors

StartTaskContact'

Fields

clientToken :: Maybe Text

A unique, case-sensitive identifier that you provide to ensure the idempotency of the request.

references :: Maybe (HashMap Text Reference)

A formatted URL that is shown to an agent in the Contact Control Panel (CCP).

previousContactId :: Maybe Text

The identifier of the previous chat, voice, or task contact.

attributes :: Maybe (HashMap Text Text)

A custom key-value pair using an attribute map. The attributes are standard Amazon Connect attributes, and can be accessed in contact flows just like any other contact attributes.

There can be up to 32,768 UTF-8 bytes across all key-value pairs per contact. Attribute keys can include only alphanumeric, dash, and underscore characters.

description :: Maybe Text

A description of the task that is shown to an agent in the Contact Control Panel (CCP).

instanceId :: Text

The identifier of the Amazon Connect instance. You can find the instanceId in the ARN of the instance.

contactFlowId :: Text

The identifier of the contact flow for initiating the tasks. To see the ContactFlowId in the Amazon Connect console user interface, on the navigation menu go to Routing, Contact Flows. Choose the contact flow. On the contact flow page, under the name of the contact flow, choose Show additional flow information. The ContactFlowId is the last part of the ARN, shown here in bold:

arn:aws:connect:us-west-2:xxxxxxxxxxxx:instance/xxxxxxxx-xxxx-xxxx-xxxx-xxxxxxxxxxxx/contact-flow/846ec553-a005-41c0-8341-xxxxxxxxxxxx

name :: Text

The name of a task that is shown to an agent in the Contact Control Panel (CCP).

newStartTaskContact

Arguments

:: Text
$sel:instanceId:StartTaskContact'

-> Text
$sel:contactFlowId:StartTaskContact'

-> Text
$sel:name:StartTaskContact'

-> StartTaskContact

Create a value of StartTaskContact with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:clientToken:StartTaskContact', startTaskContact_clientToken - A unique, case-sensitive identifier that you provide to ensure the idempotency of the request.

$sel:references:StartTaskContact', startTaskContact_references - A formatted URL that is shown to an agent in the Contact Control Panel (CCP).

$sel:previousContactId:StartTaskContact', startTaskContact_previousContactId - The identifier of the previous chat, voice, or task contact.

$sel:attributes:StartTaskContact', startTaskContact_attributes - A custom key-value pair using an attribute map. The attributes are standard Amazon Connect attributes, and can be accessed in contact flows just like any other contact attributes.

There can be up to 32,768 UTF-8 bytes across all key-value pairs per contact. Attribute keys can include only alphanumeric, dash, and underscore characters.

$sel:description:StartTaskContact', startTaskContact_description - A description of the task that is shown to an agent in the Contact Control Panel (CCP).

$sel:instanceId:StartTaskContact', startTaskContact_instanceId - The identifier of the Amazon Connect instance. You can find the instanceId in the ARN of the instance.

$sel:contactFlowId:StartTaskContact', startTaskContact_contactFlowId - The identifier of the contact flow for initiating the tasks. To see the ContactFlowId in the Amazon Connect console user interface, on the navigation menu go to Routing, Contact Flows. Choose the contact flow. On the contact flow page, under the name of the contact flow, choose Show additional flow information. The ContactFlowId is the last part of the ARN, shown here in bold:

arn:aws:connect:us-west-2:xxxxxxxxxxxx:instance/xxxxxxxx-xxxx-xxxx-xxxx-xxxxxxxxxxxx/contact-flow/846ec553-a005-41c0-8341-xxxxxxxxxxxx

$sel:name:StartTaskContact', startTaskContact_name - The name of a task that is shown to an agent in the Contact Control Panel (CCP).

Request Lenses

startTaskContact_clientToken :: Lens' StartTaskContact (Maybe Text)

A unique, case-sensitive identifier that you provide to ensure the idempotency of the request.

startTaskContact_references :: Lens' StartTaskContact (Maybe (HashMap Text Reference))

A formatted URL that is shown to an agent in the Contact Control Panel (CCP).

startTaskContact_previousContactId :: Lens' StartTaskContact (Maybe Text)

The identifier of the previous chat, voice, or task contact.

startTaskContact_attributes :: Lens' StartTaskContact (Maybe (HashMap Text Text))

A custom key-value pair using an attribute map. The attributes are standard Amazon Connect attributes, and can be accessed in contact flows just like any other contact attributes.

There can be up to 32,768 UTF-8 bytes across all key-value pairs per contact. Attribute keys can include only alphanumeric, dash, and underscore characters.

startTaskContact_description :: Lens' StartTaskContact (Maybe Text)

A description of the task that is shown to an agent in the Contact Control Panel (CCP).

startTaskContact_instanceId :: Lens' StartTaskContact Text

The identifier of the Amazon Connect instance. You can find the instanceId in the ARN of the instance.

startTaskContact_contactFlowId :: Lens' StartTaskContact Text

The identifier of the contact flow for initiating the tasks. To see the ContactFlowId in the Amazon Connect console user interface, on the navigation menu go to Routing, Contact Flows. Choose the contact flow. On the contact flow page, under the name of the contact flow, choose Show additional flow information. The ContactFlowId is the last part of the ARN, shown here in bold:

arn:aws:connect:us-west-2:xxxxxxxxxxxx:instance/xxxxxxxx-xxxx-xxxx-xxxx-xxxxxxxxxxxx/contact-flow/846ec553-a005-41c0-8341-xxxxxxxxxxxx

startTaskContact_name :: Lens' StartTaskContact Text

The name of a task that is shown to an agent in the Contact Control Panel (CCP).

Destructuring the Response

data StartTaskContactResponse

See: newStartTaskContactResponse smart constructor.

Constructors

StartTaskContactResponse'

Fields

contactId :: Maybe Text

The identifier of this contact within the Amazon Connect instance.

httpStatus :: Int

The response's http status code.

newStartTaskContactResponse

Arguments

:: Int
$sel:httpStatus:StartTaskContactResponse'

-> StartTaskContactResponse

Create a value of StartTaskContactResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:contactId:StartTaskContactResponse', startTaskContactResponse_contactId - The identifier of this contact within the Amazon Connect instance.

$sel:httpStatus:StartTaskContactResponse', startTaskContactResponse_httpStatus - The response's http status code.

Response Lenses

startTaskContactResponse_contactId :: Lens' StartTaskContactResponse (Maybe Text)

The identifier of this contact within the Amazon Connect instance.

startTaskContactResponse_httpStatus :: Lens' StartTaskContactResponse Int

The response's http status code.