Amazonka.Connect.Types.HoursOfOperationConfig

Description

Documentation

data HoursOfOperationConfig

Contains information about the hours of operation.

See: newHoursOfOperationConfig smart constructor.

Constructors

HoursOfOperationConfig'

Fields

day :: HoursOfOperationDays

The day that the hours of operation applies to.

startTime :: HoursOfOperationTimeSlice

The start time that your contact center opens.

endTime :: HoursOfOperationTimeSlice

The end time that your contact center closes.

newHoursOfOperationConfig

Arguments

:: HoursOfOperationDays
$sel:day:HoursOfOperationConfig'

-> HoursOfOperationTimeSlice
$sel:startTime:HoursOfOperationConfig'

-> HoursOfOperationTimeSlice
$sel:endTime:HoursOfOperationConfig'

-> HoursOfOperationConfig

Create a value of HoursOfOperationConfig with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:day:HoursOfOperationConfig', hoursOfOperationConfig_day - The day that the hours of operation applies to.

$sel:startTime:HoursOfOperationConfig', hoursOfOperationConfig_startTime - The start time that your contact center opens.

$sel:endTime:HoursOfOperationConfig', hoursOfOperationConfig_endTime - The end time that your contact center closes.

hoursOfOperationConfig_day :: Lens' HoursOfOperationConfig HoursOfOperationDays

The day that the hours of operation applies to.

hoursOfOperationConfig_startTime :: Lens' HoursOfOperationConfig HoursOfOperationTimeSlice

The start time that your contact center opens.

hoursOfOperationConfig_endTime :: Lens' HoursOfOperationConfig HoursOfOperationTimeSlice

The end time that your contact center closes.