Amazonka.Connect.Types.Instance

Description

Documentation

data Instance

The Amazon Connect instance.

See: newInstance smart constructor.

Constructors

Instance'

Fields

arn :: Maybe Text

The Amazon Resource Name (ARN) of the instance.

createdTime :: Maybe POSIX

When the instance was created.

outboundCallsEnabled :: Maybe Bool

Whether outbound calls are enabled.

inboundCallsEnabled :: Maybe Bool

Whether inbound calls are enabled.

instanceAlias :: Maybe (Sensitive Text)

The alias of instance.

id :: Maybe Text

The identifier of the Amazon Connect instance. You can find the instanceId in the ARN of the instance.

instanceStatus :: Maybe InstanceStatus

The state of the instance.

identityManagementType :: Maybe DirectoryType

The identity management type.

statusReason :: Maybe InstanceStatusReason

Relevant details why the instance was not successfully created.

serviceRole :: Maybe Text

The service role of the instance.

newInstance :: Instance

Create a value of Instance with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:arn:Instance', instance_arn - The Amazon Resource Name (ARN) of the instance.

$sel:createdTime:Instance', instance_createdTime - When the instance was created.

$sel:outboundCallsEnabled:Instance', instance_outboundCallsEnabled - Whether outbound calls are enabled.

$sel:inboundCallsEnabled:Instance', instance_inboundCallsEnabled - Whether inbound calls are enabled.

$sel:instanceAlias:Instance', instance_instanceAlias - The alias of instance.

$sel:id:Instance', instance_id - The identifier of the Amazon Connect instance. You can find the instanceId in the ARN of the instance.

$sel:instanceStatus:Instance', instance_instanceStatus - The state of the instance.

$sel:identityManagementType:Instance', instance_identityManagementType - The identity management type.

$sel:statusReason:Instance', instance_statusReason - Relevant details why the instance was not successfully created.

$sel:serviceRole:Instance', instance_serviceRole - The service role of the instance.

instance_arn :: Lens' Instance (Maybe Text)

The Amazon Resource Name (ARN) of the instance.

instance_createdTime :: Lens' Instance (Maybe UTCTime)

When the instance was created.

instance_outboundCallsEnabled :: Lens' Instance (Maybe Bool)

Whether outbound calls are enabled.

instance_inboundCallsEnabled :: Lens' Instance (Maybe Bool)

Whether inbound calls are enabled.

instance_instanceAlias :: Lens' Instance (Maybe Text)

The alias of instance.

instance_id :: Lens' Instance (Maybe Text)

The identifier of the Amazon Connect instance. You can find the instanceId in the ARN of the instance.

instance_instanceStatus :: Lens' Instance (Maybe InstanceStatus)

The state of the instance.

instance_identityManagementType :: Lens' Instance (Maybe DirectoryType)

The identity management type.

instance_statusReason :: Lens' Instance (Maybe InstanceStatusReason)

Relevant details why the instance was not successfully created.

instance_serviceRole :: Lens' Instance (Maybe Text)

The service role of the instance.