Amazonka.Connect.Types.Queue

Description

Documentation

data Queue

Contains information about a queue.

See: newQueue smart constructor.

Constructors

Queue'

Fields

status :: Maybe QueueStatus

The status of the queue.

queueArn :: Maybe Text

The Amazon Resource Name (ARN) for the queue.

queueId :: Maybe Text

The identifier for the queue.

maxContacts :: Maybe Natural

The maximum number of contacts that can be in the queue before it is considered full.

name :: Maybe Text

The name of the queue.

hoursOfOperationId :: Maybe Text

The identifier for the hours of operation.

outboundCallerConfig :: Maybe OutboundCallerConfig

The outbound caller ID name, number, and outbound whisper flow.

description :: Maybe Text

The description of the queue.

tags :: Maybe (HashMap Text Text)

One or more tags.

newQueue :: Queue

Create a value of Queue with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:status:Queue', queue_status - The status of the queue.

$sel:queueArn:Queue', queue_queueArn - The Amazon Resource Name (ARN) for the queue.

$sel:queueId:Queue', queue_queueId - The identifier for the queue.

$sel:maxContacts:Queue', queue_maxContacts - The maximum number of contacts that can be in the queue before it is considered full.

$sel:name:Queue', queue_name - The name of the queue.

$sel:hoursOfOperationId:Queue', queue_hoursOfOperationId - The identifier for the hours of operation.

$sel:outboundCallerConfig:Queue', queue_outboundCallerConfig - The outbound caller ID name, number, and outbound whisper flow.

$sel:description:Queue', queue_description - The description of the queue.

$sel:tags:Queue', queue_tags - One or more tags.

queue_status :: Lens' Queue (Maybe QueueStatus)

The status of the queue.

queue_queueArn :: Lens' Queue (Maybe Text)

The Amazon Resource Name (ARN) for the queue.

queue_queueId :: Lens' Queue (Maybe Text)

The identifier for the queue.

queue_maxContacts :: Lens' Queue (Maybe Natural)

The maximum number of contacts that can be in the queue before it is considered full.

queue_name :: Lens' Queue (Maybe Text)

The name of the queue.

queue_hoursOfOperationId :: Lens' Queue (Maybe Text)

The identifier for the hours of operation.

queue_outboundCallerConfig :: Lens' Queue (Maybe OutboundCallerConfig)

The outbound caller ID name, number, and outbound whisper flow.

queue_description :: Lens' Queue (Maybe Text)

The description of the queue.

queue_tags :: Lens' Queue (Maybe (HashMap Text Text))

One or more tags.