Amazonka.Connect.Types.QuickConnectConfig

Description

Documentation

data QuickConnectConfig

Contains configuration settings for a quick connect.

See: newQuickConnectConfig smart constructor.

Constructors

QuickConnectConfig'

Fields

queueConfig :: Maybe QueueQuickConnectConfig

The queue configuration. This is required only if QuickConnectType is QUEUE.

userConfig :: Maybe UserQuickConnectConfig

The user configuration. This is required only if QuickConnectType is USER.

phoneConfig :: Maybe PhoneNumberQuickConnectConfig

The phone configuration. This is required only if QuickConnectType is PHONE_NUMBER.

quickConnectType :: QuickConnectType

The type of quick connect. In the Amazon Connect console, when you create a quick connect, you are prompted to assign one of the following types: Agent (USER), External (PHONE_NUMBER), or Queue (QUEUE).

newQuickConnectConfig

Arguments

:: QuickConnectType
$sel:quickConnectType:QuickConnectConfig'

-> QuickConnectConfig

Create a value of QuickConnectConfig with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:queueConfig:QuickConnectConfig', quickConnectConfig_queueConfig - The queue configuration. This is required only if QuickConnectType is QUEUE.

$sel:userConfig:QuickConnectConfig', quickConnectConfig_userConfig - The user configuration. This is required only if QuickConnectType is USER.

$sel:phoneConfig:QuickConnectConfig', quickConnectConfig_phoneConfig - The phone configuration. This is required only if QuickConnectType is PHONE_NUMBER.

$sel:quickConnectType:QuickConnectConfig', quickConnectConfig_quickConnectType - The type of quick connect. In the Amazon Connect console, when you create a quick connect, you are prompted to assign one of the following types: Agent (USER), External (PHONE_NUMBER), or Queue (QUEUE).

quickConnectConfig_queueConfig :: Lens' QuickConnectConfig (Maybe QueueQuickConnectConfig)

The queue configuration. This is required only if QuickConnectType is QUEUE.

quickConnectConfig_userConfig :: Lens' QuickConnectConfig (Maybe UserQuickConnectConfig)

The user configuration. This is required only if QuickConnectType is USER.

quickConnectConfig_phoneConfig :: Lens' QuickConnectConfig (Maybe PhoneNumberQuickConnectConfig)

The phone configuration. This is required only if QuickConnectType is PHONE_NUMBER.

quickConnectConfig_quickConnectType :: Lens' QuickConnectConfig QuickConnectType

The type of quick connect. In the Amazon Connect console, when you create a quick connect, you are prompted to assign one of the following types: Agent (USER), External (PHONE_NUMBER), or Queue (QUEUE).