Amazonka.Connect.Types.Reference

Description

Documentation

data Reference

A link that an agent selects to complete a given task. You can have up to 4,096 UTF-8 bytes across all references for a contact.

See: newReference smart constructor.

Constructors

Reference'

Fields

value :: Text

A formatted URL that displays to an agent in the Contact Control Panel (CCP)

type' :: ReferenceType

A valid URL.

newReference

Arguments

:: Text
$sel:value:Reference'

-> ReferenceType
$sel:type':Reference'

-> Reference

Create a value of Reference with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:value:Reference', reference_value - A formatted URL that displays to an agent in the Contact Control Panel (CCP)

$sel:type':Reference', reference_type - A valid URL.

reference_value :: Lens' Reference Text

A formatted URL that displays to an agent in the Contact Control Panel (CCP)

reference_type :: Lens' Reference ReferenceType

A valid URL.