Amazonka.Connect.UpdateQueueMaxContacts

Description

This API is in preview release for Amazon Connect and is subject to change.

Updates the maximum number of contacts allowed in a queue before it is considered full.

Creating a Request

data UpdateQueueMaxContacts

See: newUpdateQueueMaxContacts smart constructor.

Constructors

UpdateQueueMaxContacts'

Fields

maxContacts :: Maybe Natural

The maximum number of contacts that can be in the queue before it is considered full.

instanceId :: Text

The identifier of the Amazon Connect instance. You can find the instanceId in the ARN of the instance.

queueId :: Text

The identifier for the queue.

newUpdateQueueMaxContacts

Arguments

:: Text
$sel:instanceId:UpdateQueueMaxContacts'

-> Text
$sel:queueId:UpdateQueueMaxContacts'

-> UpdateQueueMaxContacts

Create a value of UpdateQueueMaxContacts with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:maxContacts:UpdateQueueMaxContacts', updateQueueMaxContacts_maxContacts - The maximum number of contacts that can be in the queue before it is considered full.

$sel:instanceId:UpdateQueueMaxContacts', updateQueueMaxContacts_instanceId - The identifier of the Amazon Connect instance. You can find the instanceId in the ARN of the instance.

$sel:queueId:UpdateQueueMaxContacts', updateQueueMaxContacts_queueId - The identifier for the queue.

Request Lenses

updateQueueMaxContacts_maxContacts :: Lens' UpdateQueueMaxContacts (Maybe Natural)

The maximum number of contacts that can be in the queue before it is considered full.

updateQueueMaxContacts_instanceId :: Lens' UpdateQueueMaxContacts Text

The identifier of the Amazon Connect instance. You can find the instanceId in the ARN of the instance.

updateQueueMaxContacts_queueId :: Lens' UpdateQueueMaxContacts Text

The identifier for the queue.

Destructuring the Response

data UpdateQueueMaxContactsResponse

See: newUpdateQueueMaxContactsResponse smart constructor.

Constructors

UpdateQueueMaxContactsResponse'

newUpdateQueueMaxContactsResponse :: UpdateQueueMaxContactsResponse

Create a value of UpdateQueueMaxContactsResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.