Amazonka.Connect.UpdateQuickConnectName

Description

Updates the name and description of a quick connect. The request accepts the following data in JSON format. At least Name or Description must be provided.

Creating a Request

data UpdateQuickConnectName

See: newUpdateQuickConnectName smart constructor.

Constructors

UpdateQuickConnectName'

Fields

name :: Maybe Text

The name of the quick connect.

description :: Maybe Text

The description of the quick connect.

instanceId :: Text

The identifier of the Amazon Connect instance. You can find the instanceId in the ARN of the instance.

quickConnectId :: Text

The identifier for the quick connect.

newUpdateQuickConnectName

Arguments

:: Text
$sel:instanceId:UpdateQuickConnectName'

-> Text
$sel:quickConnectId:UpdateQuickConnectName'

-> UpdateQuickConnectName

Create a value of UpdateQuickConnectName with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:name:UpdateQuickConnectName', updateQuickConnectName_name - The name of the quick connect.

$sel:description:UpdateQuickConnectName', updateQuickConnectName_description - The description of the quick connect.

$sel:instanceId:UpdateQuickConnectName', updateQuickConnectName_instanceId - The identifier of the Amazon Connect instance. You can find the instanceId in the ARN of the instance.

$sel:quickConnectId:UpdateQuickConnectName', updateQuickConnectName_quickConnectId - The identifier for the quick connect.

Request Lenses

updateQuickConnectName_name :: Lens' UpdateQuickConnectName (Maybe Text)

The name of the quick connect.

updateQuickConnectName_description :: Lens' UpdateQuickConnectName (Maybe Text)

The description of the quick connect.

updateQuickConnectName_instanceId :: Lens' UpdateQuickConnectName Text

The identifier of the Amazon Connect instance. You can find the instanceId in the ARN of the instance.

updateQuickConnectName_quickConnectId :: Lens' UpdateQuickConnectName Text

The identifier for the quick connect.

Destructuring the Response

data UpdateQuickConnectNameResponse

See: newUpdateQuickConnectNameResponse smart constructor.

Constructors

UpdateQuickConnectNameResponse'

newUpdateQuickConnectNameResponse :: UpdateQuickConnectNameResponse

Create a value of UpdateQuickConnectNameResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.