Amazonka.Connect.UpdateUserSecurityProfiles

Description

Assigns the specified security profiles to the specified user.

Creating a Request

data UpdateUserSecurityProfiles

See: newUpdateUserSecurityProfiles smart constructor.

Constructors

UpdateUserSecurityProfiles'

Fields

securityProfileIds :: NonEmpty Text

The identifiers of the security profiles for the user.

userId :: Text

The identifier of the user account.

instanceId :: Text

The identifier of the Amazon Connect instance. You can find the instanceId in the ARN of the instance.

newUpdateUserSecurityProfiles

Arguments

:: NonEmpty Text
$sel:securityProfileIds:UpdateUserSecurityProfiles'

-> Text
$sel:userId:UpdateUserSecurityProfiles'

-> Text
$sel:instanceId:UpdateUserSecurityProfiles'

-> UpdateUserSecurityProfiles

Create a value of UpdateUserSecurityProfiles with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:securityProfileIds:UpdateUserSecurityProfiles', updateUserSecurityProfiles_securityProfileIds - The identifiers of the security profiles for the user.

$sel:userId:UpdateUserSecurityProfiles', updateUserSecurityProfiles_userId - The identifier of the user account.

$sel:instanceId:UpdateUserSecurityProfiles', updateUserSecurityProfiles_instanceId - The identifier of the Amazon Connect instance. You can find the instanceId in the ARN of the instance.

Request Lenses

updateUserSecurityProfiles_securityProfileIds :: Lens' UpdateUserSecurityProfiles (NonEmpty Text)

The identifiers of the security profiles for the user.

updateUserSecurityProfiles_userId :: Lens' UpdateUserSecurityProfiles Text

The identifier of the user account.

updateUserSecurityProfiles_instanceId :: Lens' UpdateUserSecurityProfiles Text

The identifier of the Amazon Connect instance. You can find the instanceId in the ARN of the instance.

Destructuring the Response

data UpdateUserSecurityProfilesResponse

See: newUpdateUserSecurityProfilesResponse smart constructor.

Constructors

UpdateUserSecurityProfilesResponse'

newUpdateUserSecurityProfilesResponse :: UpdateUserSecurityProfilesResponse

Create a value of UpdateUserSecurityProfilesResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.