Amazonka.ConnectContactLens

Description

Derived from API version 2020-08-21 of the AWS service descriptions, licensed under Apache 2.0.

Contact Lens for Amazon Connect enables you to analyze conversations between customer and agents, by using speech transcription, natural language processing, and intelligent search capabilities. It performs sentiment analysis, detects issues, and enables you to automatically categorize contacts.

Contact Lens for Amazon Connect provides both real-time and post-call analytics of customer-agent conversations. For more information, see Analyze conversations using Contact Lens in the Amazon Connect Administrator Guide.

Service Configuration

defaultService :: Service

API version 2020-08-21 of the Amazon Connect Contact Lens SDK configuration.

Errors

Error matchers are designed for use with the functions provided by Control.Exception.Lens. This allows catching (and rethrowing) service specific errors returned by ConnectContactLens.

AccessDeniedException

_AccessDeniedException :: AsError a => Getting (First ServiceError) a ServiceError

You do not have sufficient access to perform this action.

InvalidRequestException

_InvalidRequestException :: AsError a => Getting (First ServiceError) a ServiceError

The request is not valid.

ThrottlingException

_ThrottlingException :: AsError a => Getting (First ServiceError) a ServiceError

The throttling limit has been exceeded.

InternalServiceException

_InternalServiceException :: AsError a => Getting (First ServiceError) a ServiceError

Request processing failed due to an error or failure with the service.

ResourceNotFoundException

_ResourceNotFoundException :: AsError a => Getting (First ServiceError) a ServiceError

The specified resource was not found.

Waiters

Waiters poll by repeatedly sending a request until some remote success condition configured by the Wait specification is fulfilled. The Wait specification determines how many attempts should be made, in addition to delay and retry strategies.

Operations

Some AWS operations return results that are incomplete and require subsequent requests in order to obtain the entire result set. The process of sending subsequent requests to continue where a previous request left off is called pagination. For example, the ListObjects operation of Amazon S3 returns up to 1000 objects at a time, and you must send subsequent requests with the appropriate Marker in order to retrieve the next page of results.

Operations that have an AWSPager instance can transparently perform subsequent requests, correctly setting Markers and other request facets to iterate through the entire result set of a truncated API operation. Operations which support this have an additional note in the documentation.

Many operations have the ability to filter results on the server side. See the individual operation parameters for details.

ListRealtimeContactAnalysisSegments

data ListRealtimeContactAnalysisSegments

See: newListRealtimeContactAnalysisSegments smart constructor.

Constructors

ListRealtimeContactAnalysisSegments' (Maybe Text) (Maybe Natural) Text Text

newListRealtimeContactAnalysisSegments

Arguments

:: Text
$sel:instanceId:ListRealtimeContactAnalysisSegments'

-> Text
$sel:contactId:ListRealtimeContactAnalysisSegments'

-> ListRealtimeContactAnalysisSegments

Create a value of ListRealtimeContactAnalysisSegments with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:nextToken:ListRealtimeContactAnalysisSegments', listRealtimeContactAnalysisSegments_nextToken - The token for the next set of results. Use the value returned in the previous response in the next request to retrieve the next set of results.

$sel:maxResults:ListRealtimeContactAnalysisSegments', listRealtimeContactAnalysisSegments_maxResults - The maximimum number of results to return per page.

$sel:instanceId:ListRealtimeContactAnalysisSegments', listRealtimeContactAnalysisSegments_instanceId - The identifier of the Amazon Connect instance.

$sel:contactId:ListRealtimeContactAnalysisSegments', listRealtimeContactAnalysisSegments_contactId - The identifier of the contact.

data ListRealtimeContactAnalysisSegmentsResponse

See: newListRealtimeContactAnalysisSegmentsResponse smart constructor.

Constructors

ListRealtimeContactAnalysisSegmentsResponse' (Maybe Text) Int [RealtimeContactAnalysisSegment]

newListRealtimeContactAnalysisSegmentsResponse

Arguments

:: Int
$sel:httpStatus:ListRealtimeContactAnalysisSegmentsResponse'

-> ListRealtimeContactAnalysisSegmentsResponse

Create a value of ListRealtimeContactAnalysisSegmentsResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:nextToken:ListRealtimeContactAnalysisSegments', listRealtimeContactAnalysisSegmentsResponse_nextToken - If there are additional results, this is the token for the next set of results. If response includes nextToken there are two possible scenarios:

There are more segments so another call is required to get them.

There are no more segments at this time, but more may be available later (real-time analysis is in progress) so the client should call the operation again to get new segments.

If response does not include nextToken, the analysis is completed (successfully or failed) and there are no more segments to retrieve.

$sel:httpStatus:ListRealtimeContactAnalysisSegmentsResponse', listRealtimeContactAnalysisSegmentsResponse_httpStatus - The response's http status code.

$sel:segments:ListRealtimeContactAnalysisSegmentsResponse', listRealtimeContactAnalysisSegmentsResponse_segments - An analyzed transcript or category.

Types

SentimentValue

newtype SentimentValue

Constructors

SentimentValue'

Fields

fromSentimentValue :: Text

Bundled Patterns

pattern SentimentValue_NEGATIVE :: SentimentValue

pattern SentimentValue_NEUTRAL :: SentimentValue

pattern SentimentValue_POSITIVE :: SentimentValue

Categories

data Categories

Provides the category rules that are used to automatically categorize contacts based on uttered keywords and phrases.

See: newCategories smart constructor.

Constructors

Categories' [Text] (HashMap Text CategoryDetails)

newCategories :: Categories

Create a value of Categories with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:matchedCategories:Categories', categories_matchedCategories - The category rules that have been matched in the analyzed segment.

$sel:matchedDetails:Categories', categories_matchedDetails - The category rule that was matched and when it occurred in the transcript.

CategoryDetails

data CategoryDetails

Provides information about the category rule that was matched.

See: newCategoryDetails smart constructor.

Constructors

CategoryDetails' [PointOfInterest]

newCategoryDetails :: CategoryDetails

Create a value of CategoryDetails with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:pointsOfInterest:CategoryDetails', categoryDetails_pointsOfInterest - The section of audio where the category rule was detected.

CharacterOffsets

data CharacterOffsets

For characters that were detected as issues, where they occur in the transcript.

See: newCharacterOffsets smart constructor.

Constructors

CharacterOffsets' Natural Natural

newCharacterOffsets

Arguments

:: Natural
$sel:beginOffsetChar:CharacterOffsets'

-> Natural
$sel:endOffsetChar:CharacterOffsets'

-> CharacterOffsets

Create a value of CharacterOffsets with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:beginOffsetChar:CharacterOffsets', characterOffsets_beginOffsetChar - The beginning of the issue.

$sel:endOffsetChar:CharacterOffsets', characterOffsets_endOffsetChar - The end of the issue.

IssueDetected

data IssueDetected

Potential issues that are detected based on an artificial intelligence analysis of each turn in the conversation.

See: newIssueDetected smart constructor.

Constructors

IssueDetected' CharacterOffsets

newIssueDetected

Arguments

:: CharacterOffsets
$sel:characterOffsets:IssueDetected'

-> IssueDetected

Create a value of IssueDetected with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:characterOffsets:IssueDetected', issueDetected_characterOffsets - The offset for when the issue was detected in the segment.

PointOfInterest

data PointOfInterest

The section of the contact audio where that category rule was detected.

See: newPointOfInterest smart constructor.

Constructors

PointOfInterest' Natural Natural

newPointOfInterest

Arguments

:: Natural
$sel:beginOffsetMillis:PointOfInterest'

-> Natural
$sel:endOffsetMillis:PointOfInterest'

-> PointOfInterest

Create a value of PointOfInterest with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:beginOffsetMillis:PointOfInterest', pointOfInterest_beginOffsetMillis - The beginning offset in milliseconds where the category rule was detected.

$sel:endOffsetMillis:PointOfInterest', pointOfInterest_endOffsetMillis - The ending offset in milliseconds where the category rule was detected.

RealtimeContactAnalysisSegment

data RealtimeContactAnalysisSegment

An analyzed segment for a real-time analysis session.

See: newRealtimeContactAnalysisSegment smart constructor.

Constructors

RealtimeContactAnalysisSegment' (Maybe Categories) (Maybe Transcript)

newRealtimeContactAnalysisSegment :: RealtimeContactAnalysisSegment

Create a value of RealtimeContactAnalysisSegment with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:categories:RealtimeContactAnalysisSegment', realtimeContactAnalysisSegment_categories - The matched category rules.

$sel:transcript:RealtimeContactAnalysisSegment', realtimeContactAnalysisSegment_transcript - The analyzed transcript.

Transcript

data Transcript

A list of messages in the session.

See: newTranscript smart constructor.

Constructors

Transcript' (Maybe [IssueDetected]) Text Text Text Text Natural Natural SentimentValue

newTranscript

Arguments

:: Text
$sel:id:Transcript'

-> Text
$sel:participantId:Transcript'

-> Text
$sel:participantRole:Transcript'

-> Text
$sel:content:Transcript'

-> Natural
$sel:beginOffsetMillis:Transcript'

-> Natural
$sel:endOffsetMillis:Transcript'

-> SentimentValue
$sel:sentiment:Transcript'

-> Transcript

Create a value of Transcript with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:issuesDetected:Transcript', transcript_issuesDetected - List of positions where issues were detected on the transcript.

$sel:id:Transcript', transcript_id - The identifier of the transcript.

$sel:participantId:Transcript', transcript_participantId - The identifier of the participant.

$sel:participantRole:Transcript', transcript_participantRole - The role of participant. For example, is it a customer, agent, or system.

$sel:content:Transcript', transcript_content - The content of the transcript.

$sel:beginOffsetMillis:Transcript', transcript_beginOffsetMillis - The beginning offset in the contact for this transcript.

$sel:endOffsetMillis:Transcript', transcript_endOffsetMillis - The end offset in the contact for this transcript.

$sel:sentiment:Transcript', transcript_sentiment - The sentiment of the detected for this piece of transcript.