Amazonka.ConnectContactLens.Lens

Description

Operations

ListRealtimeContactAnalysisSegments

listRealtimeContactAnalysisSegments_nextToken :: Lens' ListRealtimeContactAnalysisSegments (Maybe Text)

The token for the next set of results. Use the value returned in the previous response in the next request to retrieve the next set of results.

listRealtimeContactAnalysisSegments_maxResults :: Lens' ListRealtimeContactAnalysisSegments (Maybe Natural)

The maximimum number of results to return per page.

listRealtimeContactAnalysisSegments_instanceId :: Lens' ListRealtimeContactAnalysisSegments Text

The identifier of the Amazon Connect instance.

listRealtimeContactAnalysisSegments_contactId :: Lens' ListRealtimeContactAnalysisSegments Text

The identifier of the contact.

listRealtimeContactAnalysisSegmentsResponse_nextToken :: Lens' ListRealtimeContactAnalysisSegmentsResponse (Maybe Text)

If there are additional results, this is the token for the next set of results. If response includes nextToken there are two possible scenarios:

There are more segments so another call is required to get them.

There are no more segments at this time, but more may be available later (real-time analysis is in progress) so the client should call the operation again to get new segments.

If response does not include nextToken, the analysis is completed (successfully or failed) and there are no more segments to retrieve.

listRealtimeContactAnalysisSegmentsResponse_httpStatus :: Lens' ListRealtimeContactAnalysisSegmentsResponse Int

The response's http status code.

listRealtimeContactAnalysisSegmentsResponse_segments :: Lens' ListRealtimeContactAnalysisSegmentsResponse [RealtimeContactAnalysisSegment]

An analyzed transcript or category.

Types

Categories

categories_matchedCategories :: Lens' Categories [Text]

The category rules that have been matched in the analyzed segment.

categories_matchedDetails :: Lens' Categories (HashMap Text CategoryDetails)

The category rule that was matched and when it occurred in the transcript.

CategoryDetails

categoryDetails_pointsOfInterest :: Lens' CategoryDetails [PointOfInterest]

The section of audio where the category rule was detected.

CharacterOffsets

characterOffsets_beginOffsetChar :: Lens' CharacterOffsets Natural

The beginning of the issue.

characterOffsets_endOffsetChar :: Lens' CharacterOffsets Natural

The end of the issue.

IssueDetected

issueDetected_characterOffsets :: Lens' IssueDetected CharacterOffsets

The offset for when the issue was detected in the segment.

PointOfInterest

pointOfInterest_beginOffsetMillis :: Lens' PointOfInterest Natural

The beginning offset in milliseconds where the category rule was detected.

pointOfInterest_endOffsetMillis :: Lens' PointOfInterest Natural

The ending offset in milliseconds where the category rule was detected.

RealtimeContactAnalysisSegment

realtimeContactAnalysisSegment_categories :: Lens' RealtimeContactAnalysisSegment (Maybe Categories)

The matched category rules.

realtimeContactAnalysisSegment_transcript :: Lens' RealtimeContactAnalysisSegment (Maybe Transcript)

The analyzed transcript.

Transcript

transcript_issuesDetected :: Lens' Transcript (Maybe [IssueDetected])

List of positions where issues were detected on the transcript.

transcript_id :: Lens' Transcript Text

The identifier of the transcript.

transcript_participantId :: Lens' Transcript Text

The identifier of the participant.

transcript_participantRole :: Lens' Transcript Text

The role of participant. For example, is it a customer, agent, or system.

transcript_content :: Lens' Transcript Text

The content of the transcript.

transcript_beginOffsetMillis :: Lens' Transcript Natural

The beginning offset in the contact for this transcript.

transcript_endOffsetMillis :: Lens' Transcript Natural

The end offset in the contact for this transcript.

transcript_sentiment :: Lens' Transcript SentimentValue

The sentiment of the detected for this piece of transcript.