Amazonka.ConnectParticipant

Description

Derived from API version 2018-09-07 of the AWS service descriptions, licensed under Apache 2.0.

Amazon Connect is a cloud-based contact center solution that makes it easy to set up and manage a customer contact center and provide reliable customer engagement at any scale.

Amazon Connect enables customer contacts through voice or chat.

The APIs described here are used by chat participants, such as agents and customers.

Service Configuration

defaultService :: Service

API version 2018-09-07 of the Amazon Connect Participant Service SDK configuration.

Errors

Error matchers are designed for use with the functions provided by Control.Exception.Lens. This allows catching (and rethrowing) service specific errors returned by ConnectParticipant.

ValidationException

_ValidationException :: AsError a => Getting (First ServiceError) a ServiceError

The input fails to satisfy the constraints specified by Amazon Connect.

AccessDeniedException

_AccessDeniedException :: AsError a => Getting (First ServiceError) a ServiceError

You do not have sufficient access to perform this action.

ConflictException

_ConflictException :: AsError a => Getting (First ServiceError) a ServiceError

An attachment with that identifier is already being uploaded.

ServiceQuotaExceededException

_ServiceQuotaExceededException :: AsError a => Getting (First ServiceError) a ServiceError

The number of attachments per contact exceeds the quota.

ThrottlingException

_ThrottlingException :: AsError a => Getting (First ServiceError) a ServiceError

The request was denied due to request throttling.

InternalServerException

_InternalServerException :: AsError a => Getting (First ServiceError) a ServiceError

This exception occurs when there is an internal failure in the Amazon Connect service.

Waiters

Waiters poll by repeatedly sending a request until some remote success condition configured by the Wait specification is fulfilled. The Wait specification determines how many attempts should be made, in addition to delay and retry strategies.

Operations

Some AWS operations return results that are incomplete and require subsequent requests in order to obtain the entire result set. The process of sending subsequent requests to continue where a previous request left off is called pagination. For example, the ListObjects operation of Amazon S3 returns up to 1000 objects at a time, and you must send subsequent requests with the appropriate Marker in order to retrieve the next page of results.

Operations that have an AWSPager instance can transparently perform subsequent requests, correctly setting Markers and other request facets to iterate through the entire result set of a truncated API operation. Operations which support this have an additional note in the documentation.

Many operations have the ability to filter results on the server side. See the individual operation parameters for details.

SendMessage

data SendMessage

See: newSendMessage smart constructor.

Constructors

SendMessage' (Maybe Text) Text Text Text

newSendMessage

Arguments

:: Text
$sel:contentType:SendMessage'

-> Text
$sel:content:SendMessage'

-> Text
$sel:connectionToken:SendMessage'

-> SendMessage

Create a value of SendMessage with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:clientToken:SendMessage', sendMessage_clientToken - A unique, case-sensitive identifier that you provide to ensure the idempotency of the request.

$sel:contentType:SendMessage', sendMessage_contentType - The type of the content. Supported types are text/plain.

$sel:content:SendMessage', sendMessage_content - The content of the message.

$sel:connectionToken:SendMessage', sendMessage_connectionToken - The authentication token associated with the connection.

data SendMessageResponse

See: newSendMessageResponse smart constructor.

Constructors

SendMessageResponse' (Maybe Text) (Maybe Text) Int

newSendMessageResponse

Arguments

:: Int
$sel:httpStatus:SendMessageResponse'

-> SendMessageResponse

Create a value of SendMessageResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:absoluteTime:SendMessageResponse', sendMessageResponse_absoluteTime - The time when the message was sent.

It's specified in ISO 8601 format: yyyy-MM-ddThh:mm:ss.SSSZ. For example, 2019-11-08T02:41:28.172Z.

$sel:id:SendMessageResponse', sendMessageResponse_id - The ID of the message.

$sel:httpStatus:SendMessageResponse', sendMessageResponse_httpStatus - The response's http status code.

DisconnectParticipant

data DisconnectParticipant

See: newDisconnectParticipant smart constructor.

Constructors

DisconnectParticipant' (Maybe Text) Text

newDisconnectParticipant

Arguments

:: Text
$sel:connectionToken:DisconnectParticipant'

-> DisconnectParticipant

Create a value of DisconnectParticipant with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:clientToken:DisconnectParticipant', disconnectParticipant_clientToken - A unique, case-sensitive identifier that you provide to ensure the idempotency of the request.

$sel:connectionToken:DisconnectParticipant', disconnectParticipant_connectionToken - The authentication token associated with the participant's connection.

data DisconnectParticipantResponse

See: newDisconnectParticipantResponse smart constructor.

Constructors

DisconnectParticipantResponse' Int

newDisconnectParticipantResponse

Arguments

:: Int
$sel:httpStatus:DisconnectParticipantResponse'

-> DisconnectParticipantResponse

Create a value of DisconnectParticipantResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:httpStatus:DisconnectParticipantResponse', disconnectParticipantResponse_httpStatus - The response's http status code.

CompleteAttachmentUpload

data CompleteAttachmentUpload

See: newCompleteAttachmentUpload smart constructor.

Constructors

CompleteAttachmentUpload' (NonEmpty Text) Text Text

newCompleteAttachmentUpload

Arguments

:: NonEmpty Text
$sel:attachmentIds:CompleteAttachmentUpload'

-> Text
$sel:clientToken:CompleteAttachmentUpload'

-> Text
$sel:connectionToken:CompleteAttachmentUpload'

-> CompleteAttachmentUpload

Create a value of CompleteAttachmentUpload with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:attachmentIds:CompleteAttachmentUpload', completeAttachmentUpload_attachmentIds - A list of unique identifiers for the attachments.

$sel:clientToken:CompleteAttachmentUpload', completeAttachmentUpload_clientToken - A unique, case-sensitive identifier that you provide to ensure the idempotency of the request.

$sel:connectionToken:CompleteAttachmentUpload', completeAttachmentUpload_connectionToken - The authentication token associated with the participant's connection.

data CompleteAttachmentUploadResponse

See: newCompleteAttachmentUploadResponse smart constructor.

Constructors

CompleteAttachmentUploadResponse' Int

newCompleteAttachmentUploadResponse

Arguments

:: Int
$sel:httpStatus:CompleteAttachmentUploadResponse'

-> CompleteAttachmentUploadResponse

Create a value of CompleteAttachmentUploadResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:httpStatus:CompleteAttachmentUploadResponse', completeAttachmentUploadResponse_httpStatus - The response's http status code.

GetAttachment

data GetAttachment

See: newGetAttachment smart constructor.

Constructors

GetAttachment' Text Text

newGetAttachment

Arguments

:: Text
$sel:attachmentId:GetAttachment'

-> Text
$sel:connectionToken:GetAttachment'

-> GetAttachment

Create a value of GetAttachment with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:attachmentId:GetAttachment', getAttachment_attachmentId - A unique identifier for the attachment.

$sel:connectionToken:GetAttachment', getAttachment_connectionToken - The authentication token associated with the participant's connection.

data GetAttachmentResponse

See: newGetAttachmentResponse smart constructor.

Constructors

GetAttachmentResponse' (Maybe Text) (Maybe Text) Int

newGetAttachmentResponse

Arguments

:: Int
$sel:httpStatus:GetAttachmentResponse'

-> GetAttachmentResponse

Create a value of GetAttachmentResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:urlExpiry:GetAttachmentResponse', getAttachmentResponse_urlExpiry - The expiration time of the URL in ISO timestamp. It's specified in ISO 8601 format: yyyy-MM-ddThh:mm:ss.SSSZ. For example, 2019-11-08T02:41:28.172Z.

$sel:url:GetAttachmentResponse', getAttachmentResponse_url - The pre-signed URL using which file would be downloaded from Amazon S3 by the API caller.

$sel:httpStatus:GetAttachmentResponse', getAttachmentResponse_httpStatus - The response's http status code.

StartAttachmentUpload

data StartAttachmentUpload

See: newStartAttachmentUpload smart constructor.

Constructors

StartAttachmentUpload' Text Natural Text Text Text

newStartAttachmentUpload

Arguments

:: Text
$sel:contentType:StartAttachmentUpload'

-> Natural
$sel:attachmentSizeInBytes:StartAttachmentUpload'

-> Text
$sel:attachmentName:StartAttachmentUpload'

-> Text
$sel:clientToken:StartAttachmentUpload'

-> Text
$sel:connectionToken:StartAttachmentUpload'

-> StartAttachmentUpload

Create a value of StartAttachmentUpload with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:contentType:StartAttachmentUpload', startAttachmentUpload_contentType - Describes the MIME file type of the attachment. For a list of supported file types, see Feature specifications in the Amazon Connect Administrator Guide.

$sel:attachmentSizeInBytes:StartAttachmentUpload', startAttachmentUpload_attachmentSizeInBytes - The size of the attachment in bytes.

$sel:attachmentName:StartAttachmentUpload', startAttachmentUpload_attachmentName - A case-sensitive name of the attachment being uploaded.

$sel:clientToken:StartAttachmentUpload', startAttachmentUpload_clientToken - A unique case sensitive identifier to support idempotency of request.

$sel:connectionToken:StartAttachmentUpload', startAttachmentUpload_connectionToken - The authentication token associated with the participant's connection.

data StartAttachmentUploadResponse

See: newStartAttachmentUploadResponse smart constructor.

Constructors

StartAttachmentUploadResponse' (Maybe Text) (Maybe UploadMetadata) Int

newStartAttachmentUploadResponse

Arguments

:: Int
$sel:httpStatus:StartAttachmentUploadResponse'

-> StartAttachmentUploadResponse

Create a value of StartAttachmentUploadResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:attachmentId:StartAttachmentUploadResponse', startAttachmentUploadResponse_attachmentId - A unique identifier for the attachment.

$sel:uploadMetadata:StartAttachmentUploadResponse', startAttachmentUploadResponse_uploadMetadata - Fields to be used while uploading the attachment.

$sel:httpStatus:StartAttachmentUploadResponse', startAttachmentUploadResponse_httpStatus - The response's http status code.

CreateParticipantConnection

data CreateParticipantConnection

See: newCreateParticipantConnection smart constructor.

Constructors

CreateParticipantConnection' (NonEmpty ConnectionType) Text

newCreateParticipantConnection

Arguments

:: NonEmpty ConnectionType
$sel:type':CreateParticipantConnection'

-> Text
$sel:participantToken:CreateParticipantConnection'

-> CreateParticipantConnection

Create a value of CreateParticipantConnection with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:type':CreateParticipantConnection', createParticipantConnection_type - Type of connection information required.

$sel:participantToken:CreateParticipantConnection', createParticipantConnection_participantToken - This is a header parameter.

The Participant Token as obtained from StartChatContact API response.

data CreateParticipantConnectionResponse

See: newCreateParticipantConnectionResponse smart constructor.

Constructors

CreateParticipantConnectionResponse' (Maybe ConnectionCredentials) (Maybe Websocket) Int

newCreateParticipantConnectionResponse

Arguments

:: Int
$sel:httpStatus:CreateParticipantConnectionResponse'

-> CreateParticipantConnectionResponse

Create a value of CreateParticipantConnectionResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:connectionCredentials:CreateParticipantConnectionResponse', createParticipantConnectionResponse_connectionCredentials - Creates the participant's connection credentials. The authentication token associated with the participant's connection.

$sel:websocket:CreateParticipantConnectionResponse', createParticipantConnectionResponse_websocket - Creates the participant's websocket connection.

$sel:httpStatus:CreateParticipantConnectionResponse', createParticipantConnectionResponse_httpStatus - The response's http status code.

GetTranscript

data GetTranscript

See: newGetTranscript smart constructor.

Constructors

GetTranscript' (Maybe ScanDirection) (Maybe Text) (Maybe SortKey) (Maybe Text) (Maybe Natural) (Maybe StartPosition) Text

newGetTranscript

Arguments

:: Text
$sel:connectionToken:GetTranscript'

-> GetTranscript

Create a value of GetTranscript with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:scanDirection:GetTranscript', getTranscript_scanDirection - The direction from StartPosition from which to retrieve message. Default: BACKWARD when no StartPosition is provided, FORWARD with StartPosition.

$sel:nextToken:GetTranscript', getTranscript_nextToken - The pagination token. Use the value returned previously in the next subsequent request to retrieve the next set of results.

$sel:sortOrder:GetTranscript', getTranscript_sortOrder - The sort order for the records. Default: DESCENDING.

$sel:contactId:GetTranscript', getTranscript_contactId - The contactId from the current contact chain for which transcript is needed.

$sel:maxResults:GetTranscript', getTranscript_maxResults - The maximum number of results to return in the page. Default: 10.

$sel:startPosition:GetTranscript', getTranscript_startPosition - A filtering option for where to start.

$sel:connectionToken:GetTranscript', getTranscript_connectionToken - The authentication token associated with the participant's connection.

data GetTranscriptResponse

See: newGetTranscriptResponse smart constructor.

Constructors

GetTranscriptResponse' (Maybe Text) (Maybe Text) (Maybe [Item]) Int

newGetTranscriptResponse

Arguments

:: Int
$sel:httpStatus:GetTranscriptResponse'

-> GetTranscriptResponse

Create a value of GetTranscriptResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:nextToken:GetTranscript', getTranscriptResponse_nextToken - The pagination token. Use the value returned previously in the next subsequent request to retrieve the next set of results.

$sel:initialContactId:GetTranscriptResponse', getTranscriptResponse_initialContactId - The initial contact ID for the contact.

$sel:transcript:GetTranscriptResponse', getTranscriptResponse_transcript - The list of messages in the session.

$sel:httpStatus:GetTranscriptResponse', getTranscriptResponse_httpStatus - The response's http status code.

SendEvent

data SendEvent

See: newSendEvent smart constructor.

Constructors

SendEvent' (Maybe Text) (Maybe Text) Text Text

newSendEvent

Arguments

:: Text
$sel:contentType:SendEvent'

-> Text
$sel:connectionToken:SendEvent'

-> SendEvent

Create a value of SendEvent with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:clientToken:SendEvent', sendEvent_clientToken - A unique, case-sensitive identifier that you provide to ensure the idempotency of the request.

$sel:content:SendEvent', sendEvent_content - The content of the event to be sent (for example, message text). This is not yet supported.

$sel:contentType:SendEvent', sendEvent_contentType - The content type of the request. Supported types are:

application/vnd.amazonaws.connect.event.typing

application/vnd.amazonaws.connect.event.connection.acknowledged

$sel:connectionToken:SendEvent', sendEvent_connectionToken - The authentication token associated with the participant's connection.

data SendEventResponse

See: newSendEventResponse smart constructor.

Constructors

SendEventResponse' (Maybe Text) (Maybe Text) Int

newSendEventResponse

Arguments

:: Int
$sel:httpStatus:SendEventResponse'

-> SendEventResponse

Create a value of SendEventResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:absoluteTime:SendEventResponse', sendEventResponse_absoluteTime - The time when the event was sent.

It's specified in ISO 8601 format: yyyy-MM-ddThh:mm:ss.SSSZ. For example, 2019-11-08T02:41:28.172Z.

$sel:id:SendEventResponse', sendEventResponse_id - The ID of the response.

$sel:httpStatus:SendEventResponse', sendEventResponse_httpStatus - The response's http status code.

Types

ArtifactStatus

newtype ArtifactStatus

Constructors

ArtifactStatus'

Fields

fromArtifactStatus :: Text

Bundled Patterns

pattern ArtifactStatus_APPROVED :: ArtifactStatus

pattern ArtifactStatus_IN_PROGRESS :: ArtifactStatus

pattern ArtifactStatus_REJECTED :: ArtifactStatus

ChatItemType

newtype ChatItemType

Constructors

ChatItemType'

Fields

fromChatItemType :: Text

Bundled Patterns

pattern ChatItemType_ATTACHMENT :: ChatItemType

pattern ChatItemType_CHAT_ENDED :: ChatItemType

pattern ChatItemType_CONNECTION_ACK :: ChatItemType

pattern ChatItemType_EVENT :: ChatItemType

pattern ChatItemType_MESSAGE :: ChatItemType

pattern ChatItemType_PARTICIPANT_JOINED :: ChatItemType

pattern ChatItemType_PARTICIPANT_LEFT :: ChatItemType

pattern ChatItemType_TRANSFER_FAILED :: ChatItemType

pattern ChatItemType_TRANSFER_SUCCEEDED :: ChatItemType

pattern ChatItemType_TYPING :: ChatItemType

ConnectionType

newtype ConnectionType

Constructors

ConnectionType'

Fields

fromConnectionType :: Text

Bundled Patterns

pattern ConnectionType_CONNECTION_CREDENTIALS :: ConnectionType

pattern ConnectionType_WEBSOCKET :: ConnectionType

ParticipantRole

newtype ParticipantRole

Constructors

ParticipantRole'

Fields

fromParticipantRole :: Text

Bundled Patterns

pattern ParticipantRole_AGENT :: ParticipantRole

pattern ParticipantRole_CUSTOMER :: ParticipantRole

pattern ParticipantRole_SYSTEM :: ParticipantRole

ScanDirection

newtype ScanDirection

Constructors

ScanDirection'

Fields

fromScanDirection :: Text

Bundled Patterns

pattern ScanDirection_BACKWARD :: ScanDirection

pattern ScanDirection_FORWARD :: ScanDirection

SortKey

newtype SortKey

Constructors

SortKey'

Fields

fromSortKey :: Text

Bundled Patterns

pattern SortKey_ASCENDING :: SortKey

pattern SortKey_DESCENDING :: SortKey

AttachmentItem

data AttachmentItem

The case-insensitive input to indicate standard MIME type that describes the format of the file that will be uploaded.

See: newAttachmentItem smart constructor.

Constructors

AttachmentItem' (Maybe ArtifactStatus) (Maybe Text) (Maybe Text) (Maybe Text)

newAttachmentItem :: AttachmentItem

Create a value of AttachmentItem with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:status:AttachmentItem', attachmentItem_status - Status of the attachment.

$sel:attachmentName:AttachmentItem', attachmentItem_attachmentName - A case-sensitive name of the attachment being uploaded.

$sel:attachmentId:AttachmentItem', attachmentItem_attachmentId - A unique identifier for the attachment.

$sel:contentType:AttachmentItem', attachmentItem_contentType - Describes the MIME file type of the attachment. For a list of supported file types, see Feature specifications in the Amazon Connect Administrator Guide.

ConnectionCredentials

data ConnectionCredentials

Connection credentials.

See: newConnectionCredentials smart constructor.

Constructors

ConnectionCredentials' (Maybe Text) (Maybe Text)

newConnectionCredentials :: ConnectionCredentials

Create a value of ConnectionCredentials with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:expiry:ConnectionCredentials', connectionCredentials_expiry - The expiration of the token.

It's specified in ISO 8601 format: yyyy-MM-ddThh:mm:ss.SSSZ. For example, 2019-11-08T02:41:28.172Z.

$sel:connectionToken:ConnectionCredentials', connectionCredentials_connectionToken - The connection token.

Item

data Item

An item - message or event - that has been sent.

See: newItem smart constructor.

Constructors

Item' (Maybe Text) (Maybe Text) (Maybe [AttachmentItem]) (Maybe ParticipantRole) (Maybe Text) (Maybe Text) (Maybe Text) (Maybe ChatItemType) (Maybe Text)

newItem :: Item

Create a value of Item with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:participantId:Item', item_participantId - The ID of the sender in the session.

$sel:absoluteTime:Item', item_absoluteTime - The time when the message or event was sent.

It's specified in ISO 8601 format: yyyy-MM-ddThh:mm:ss.SSSZ. For example, 2019-11-08T02:41:28.172Z.

$sel:attachments:Item', item_attachments - Provides information about the attachments.

$sel:participantRole:Item', item_participantRole - The role of the sender. For example, is it a customer, agent, or system.

$sel:content:Item', item_content - The content of the message or event.

$sel:id:Item', item_id - The ID of the item.

$sel:displayName:Item', item_displayName - The chat display name of the sender.

$sel:type':Item', item_type - Type of the item: message or event.

$sel:contentType:Item', item_contentType - The type of content of the item.

StartPosition

data StartPosition

A filtering option for where to start. For example, if you sent 100 messages, start with message 50.

See: newStartPosition smart constructor.

Constructors

StartPosition' (Maybe Text) (Maybe Text) (Maybe Natural)

newStartPosition :: StartPosition

Create a value of StartPosition with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:absoluteTime:StartPosition', startPosition_absoluteTime - The time in ISO format where to start.

It's specified in ISO 8601 format: yyyy-MM-ddThh:mm:ss.SSSZ. For example, 2019-11-08T02:41:28.172Z.

$sel:id:StartPosition', startPosition_id - The ID of the message or event where to start.

$sel:mostRecent:StartPosition', startPosition_mostRecent - The start position of the most recent message where you want to start.

UploadMetadata

data UploadMetadata

Fields to be used while uploading the attachment.

See: newUploadMetadata smart constructor.

Constructors

UploadMetadata' (Maybe Text) (Maybe (HashMap Text Text)) (Maybe Text)

newUploadMetadata :: UploadMetadata

Create a value of UploadMetadata with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:urlExpiry:UploadMetadata', uploadMetadata_urlExpiry - The expiration time of the URL in ISO timestamp. It's specified in ISO 8601 format: yyyy-MM-ddThh:mm:ss.SSSZ. For example, 2019-11-08T02:41:28.172Z.

$sel:headersToInclude:UploadMetadata', uploadMetadata_headersToInclude - The headers to be provided while uploading the file to the URL.

$sel:url:UploadMetadata', uploadMetadata_url - The pre-signed URL using which file would be downloaded from Amazon S3 by the API caller.

Websocket

data Websocket

The websocket for the participant's connection.

See: newWebsocket smart constructor.

Constructors

Websocket' (Maybe Text) (Maybe Text)

newWebsocket :: Websocket

Create a value of Websocket with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:url:Websocket', websocket_url - The URL of the websocket.

$sel:connectionExpiry:Websocket', websocket_connectionExpiry - The URL expiration timestamp in ISO date format.

It's specified in ISO 8601 format: yyyy-MM-ddThh:mm:ss.SSSZ. For example, 2019-11-08T02:41:28.172Z.