Amazonka.ConnectParticipant.GetTranscript

Description

Retrieves a transcript of the session, including details about any attachments. Note that ConnectionToken is used for invoking this API instead of ParticipantToken.

The Amazon Connect Participant Service APIs do not use Signature Version 4 authentication.

Creating a Request

data GetTranscript

See: newGetTranscript smart constructor.

Constructors

GetTranscript'

Fields

scanDirection :: Maybe ScanDirection

The direction from StartPosition from which to retrieve message. Default: BACKWARD when no StartPosition is provided, FORWARD with StartPosition.

nextToken :: Maybe Text

The pagination token. Use the value returned previously in the next subsequent request to retrieve the next set of results.

sortOrder :: Maybe SortKey

The sort order for the records. Default: DESCENDING.

contactId :: Maybe Text

The contactId from the current contact chain for which transcript is needed.

maxResults :: Maybe Natural

The maximum number of results to return in the page. Default: 10.

startPosition :: Maybe StartPosition

A filtering option for where to start.

connectionToken :: Text

The authentication token associated with the participant's connection.

newGetTranscript

Arguments

:: Text
$sel:connectionToken:GetTranscript'

-> GetTranscript

Create a value of GetTranscript with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:scanDirection:GetTranscript', getTranscript_scanDirection - The direction from StartPosition from which to retrieve message. Default: BACKWARD when no StartPosition is provided, FORWARD with StartPosition.

$sel:nextToken:GetTranscript', getTranscript_nextToken - The pagination token. Use the value returned previously in the next subsequent request to retrieve the next set of results.

$sel:sortOrder:GetTranscript', getTranscript_sortOrder - The sort order for the records. Default: DESCENDING.

$sel:contactId:GetTranscript', getTranscript_contactId - The contactId from the current contact chain for which transcript is needed.

$sel:maxResults:GetTranscript', getTranscript_maxResults - The maximum number of results to return in the page. Default: 10.

$sel:startPosition:GetTranscript', getTranscript_startPosition - A filtering option for where to start.

$sel:connectionToken:GetTranscript', getTranscript_connectionToken - The authentication token associated with the participant's connection.

Request Lenses

getTranscript_scanDirection :: Lens' GetTranscript (Maybe ScanDirection)

The direction from StartPosition from which to retrieve message. Default: BACKWARD when no StartPosition is provided, FORWARD with StartPosition.

getTranscript_nextToken :: Lens' GetTranscript (Maybe Text)

The pagination token. Use the value returned previously in the next subsequent request to retrieve the next set of results.

getTranscript_sortOrder :: Lens' GetTranscript (Maybe SortKey)

The sort order for the records. Default: DESCENDING.

getTranscript_contactId :: Lens' GetTranscript (Maybe Text)

The contactId from the current contact chain for which transcript is needed.

getTranscript_maxResults :: Lens' GetTranscript (Maybe Natural)

The maximum number of results to return in the page. Default: 10.

getTranscript_startPosition :: Lens' GetTranscript (Maybe StartPosition)

A filtering option for where to start.

getTranscript_connectionToken :: Lens' GetTranscript Text

The authentication token associated with the participant's connection.

Destructuring the Response

data GetTranscriptResponse

See: newGetTranscriptResponse smart constructor.

Constructors

GetTranscriptResponse'

Fields

nextToken :: Maybe Text

The pagination token. Use the value returned previously in the next subsequent request to retrieve the next set of results.

initialContactId :: Maybe Text

The initial contact ID for the contact.

transcript :: Maybe [Item]

The list of messages in the session.

httpStatus :: Int

The response's http status code.

newGetTranscriptResponse

Arguments

:: Int
$sel:httpStatus:GetTranscriptResponse'

-> GetTranscriptResponse

Create a value of GetTranscriptResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:nextToken:GetTranscript', getTranscriptResponse_nextToken - The pagination token. Use the value returned previously in the next subsequent request to retrieve the next set of results.

$sel:initialContactId:GetTranscriptResponse', getTranscriptResponse_initialContactId - The initial contact ID for the contact.

$sel:transcript:GetTranscriptResponse', getTranscriptResponse_transcript - The list of messages in the session.

$sel:httpStatus:GetTranscriptResponse', getTranscriptResponse_httpStatus - The response's http status code.

Response Lenses

getTranscriptResponse_nextToken :: Lens' GetTranscriptResponse (Maybe Text)

The pagination token. Use the value returned previously in the next subsequent request to retrieve the next set of results.

getTranscriptResponse_initialContactId :: Lens' GetTranscriptResponse (Maybe Text)

The initial contact ID for the contact.

getTranscriptResponse_transcript :: Lens' GetTranscriptResponse (Maybe [Item])

The list of messages in the session.

getTranscriptResponse_httpStatus :: Lens' GetTranscriptResponse Int

The response's http status code.