Amazonka.ConnectParticipant.Lens

Description

Operations

SendMessage

sendMessage_clientToken :: Lens' SendMessage (Maybe Text)

A unique, case-sensitive identifier that you provide to ensure the idempotency of the request.

sendMessage_contentType :: Lens' SendMessage Text

The type of the content. Supported types are text/plain.

sendMessage_content :: Lens' SendMessage Text

The content of the message.

sendMessage_connectionToken :: Lens' SendMessage Text

The authentication token associated with the connection.

sendMessageResponse_absoluteTime :: Lens' SendMessageResponse (Maybe Text)

The time when the message was sent.

It's specified in ISO 8601 format: yyyy-MM-ddThh:mm:ss.SSSZ. For example, 2019-11-08T02:41:28.172Z.

sendMessageResponse_id :: Lens' SendMessageResponse (Maybe Text)

The ID of the message.

sendMessageResponse_httpStatus :: Lens' SendMessageResponse Int

The response's http status code.

DisconnectParticipant

disconnectParticipant_clientToken :: Lens' DisconnectParticipant (Maybe Text)

A unique, case-sensitive identifier that you provide to ensure the idempotency of the request.

disconnectParticipant_connectionToken :: Lens' DisconnectParticipant Text

The authentication token associated with the participant's connection.

disconnectParticipantResponse_httpStatus :: Lens' DisconnectParticipantResponse Int

The response's http status code.

CompleteAttachmentUpload

completeAttachmentUpload_attachmentIds :: Lens' CompleteAttachmentUpload (NonEmpty Text)

A list of unique identifiers for the attachments.

completeAttachmentUpload_clientToken :: Lens' CompleteAttachmentUpload Text

A unique, case-sensitive identifier that you provide to ensure the idempotency of the request.

completeAttachmentUpload_connectionToken :: Lens' CompleteAttachmentUpload Text

The authentication token associated with the participant's connection.

completeAttachmentUploadResponse_httpStatus :: Lens' CompleteAttachmentUploadResponse Int

The response's http status code.

GetAttachment

getAttachment_attachmentId :: Lens' GetAttachment Text

A unique identifier for the attachment.

getAttachment_connectionToken :: Lens' GetAttachment Text

The authentication token associated with the participant's connection.

getAttachmentResponse_urlExpiry :: Lens' GetAttachmentResponse (Maybe Text)

The expiration time of the URL in ISO timestamp. It's specified in ISO 8601 format: yyyy-MM-ddThh:mm:ss.SSSZ. For example, 2019-11-08T02:41:28.172Z.

getAttachmentResponse_url :: Lens' GetAttachmentResponse (Maybe Text)

The pre-signed URL using which file would be downloaded from Amazon S3 by the API caller.

getAttachmentResponse_httpStatus :: Lens' GetAttachmentResponse Int

The response's http status code.

StartAttachmentUpload

startAttachmentUpload_contentType :: Lens' StartAttachmentUpload Text

Describes the MIME file type of the attachment. For a list of supported file types, see Feature specifications in the Amazon Connect Administrator Guide.

startAttachmentUpload_attachmentSizeInBytes :: Lens' StartAttachmentUpload Natural

The size of the attachment in bytes.

startAttachmentUpload_attachmentName :: Lens' StartAttachmentUpload Text

A case-sensitive name of the attachment being uploaded.

startAttachmentUpload_clientToken :: Lens' StartAttachmentUpload Text

A unique case sensitive identifier to support idempotency of request.

startAttachmentUpload_connectionToken :: Lens' StartAttachmentUpload Text

The authentication token associated with the participant's connection.

startAttachmentUploadResponse_attachmentId :: Lens' StartAttachmentUploadResponse (Maybe Text)

A unique identifier for the attachment.

startAttachmentUploadResponse_uploadMetadata :: Lens' StartAttachmentUploadResponse (Maybe UploadMetadata)

Fields to be used while uploading the attachment.

startAttachmentUploadResponse_httpStatus :: Lens' StartAttachmentUploadResponse Int

The response's http status code.

CreateParticipantConnection

createParticipantConnection_type :: Lens' CreateParticipantConnection (NonEmpty ConnectionType)

Type of connection information required.

createParticipantConnection_participantToken :: Lens' CreateParticipantConnection Text

This is a header parameter.

The Participant Token as obtained from StartChatContact API response.

createParticipantConnectionResponse_connectionCredentials :: Lens' CreateParticipantConnectionResponse (Maybe ConnectionCredentials)

Creates the participant's connection credentials. The authentication token associated with the participant's connection.

createParticipantConnectionResponse_websocket :: Lens' CreateParticipantConnectionResponse (Maybe Websocket)

Creates the participant's websocket connection.

createParticipantConnectionResponse_httpStatus :: Lens' CreateParticipantConnectionResponse Int

The response's http status code.

GetTranscript

getTranscript_scanDirection :: Lens' GetTranscript (Maybe ScanDirection)

The direction from StartPosition from which to retrieve message. Default: BACKWARD when no StartPosition is provided, FORWARD with StartPosition.

getTranscript_nextToken :: Lens' GetTranscript (Maybe Text)

The pagination token. Use the value returned previously in the next subsequent request to retrieve the next set of results.

getTranscript_sortOrder :: Lens' GetTranscript (Maybe SortKey)

The sort order for the records. Default: DESCENDING.

getTranscript_contactId :: Lens' GetTranscript (Maybe Text)

The contactId from the current contact chain for which transcript is needed.

getTranscript_maxResults :: Lens' GetTranscript (Maybe Natural)

The maximum number of results to return in the page. Default: 10.

getTranscript_startPosition :: Lens' GetTranscript (Maybe StartPosition)

A filtering option for where to start.

getTranscript_connectionToken :: Lens' GetTranscript Text

The authentication token associated with the participant's connection.

getTranscriptResponse_nextToken :: Lens' GetTranscriptResponse (Maybe Text)

The pagination token. Use the value returned previously in the next subsequent request to retrieve the next set of results.

getTranscriptResponse_initialContactId :: Lens' GetTranscriptResponse (Maybe Text)

The initial contact ID for the contact.

getTranscriptResponse_transcript :: Lens' GetTranscriptResponse (Maybe [Item])

The list of messages in the session.

getTranscriptResponse_httpStatus :: Lens' GetTranscriptResponse Int

The response's http status code.

SendEvent

sendEvent_clientToken :: Lens' SendEvent (Maybe Text)

A unique, case-sensitive identifier that you provide to ensure the idempotency of the request.

sendEvent_content :: Lens' SendEvent (Maybe Text)

The content of the event to be sent (for example, message text). This is not yet supported.

sendEvent_contentType :: Lens' SendEvent Text

The content type of the request. Supported types are:

application/vnd.amazonaws.connect.event.typing

application/vnd.amazonaws.connect.event.connection.acknowledged

sendEvent_connectionToken :: Lens' SendEvent Text

The authentication token associated with the participant's connection.

sendEventResponse_absoluteTime :: Lens' SendEventResponse (Maybe Text)

The time when the event was sent.

It's specified in ISO 8601 format: yyyy-MM-ddThh:mm:ss.SSSZ. For example, 2019-11-08T02:41:28.172Z.

sendEventResponse_id :: Lens' SendEventResponse (Maybe Text)

The ID of the response.

sendEventResponse_httpStatus :: Lens' SendEventResponse Int

The response's http status code.

Types

AttachmentItem

attachmentItem_status :: Lens' AttachmentItem (Maybe ArtifactStatus)

Status of the attachment.

attachmentItem_attachmentName :: Lens' AttachmentItem (Maybe Text)

A case-sensitive name of the attachment being uploaded.

attachmentItem_attachmentId :: Lens' AttachmentItem (Maybe Text)

A unique identifier for the attachment.

attachmentItem_contentType :: Lens' AttachmentItem (Maybe Text)

Describes the MIME file type of the attachment. For a list of supported file types, see Feature specifications in the Amazon Connect Administrator Guide.

ConnectionCredentials

connectionCredentials_expiry :: Lens' ConnectionCredentials (Maybe Text)

The expiration of the token.

It's specified in ISO 8601 format: yyyy-MM-ddThh:mm:ss.SSSZ. For example, 2019-11-08T02:41:28.172Z.

connectionCredentials_connectionToken :: Lens' ConnectionCredentials (Maybe Text)

The connection token.

Item

item_participantId :: Lens' Item (Maybe Text)

The ID of the sender in the session.

item_absoluteTime :: Lens' Item (Maybe Text)

The time when the message or event was sent.

It's specified in ISO 8601 format: yyyy-MM-ddThh:mm:ss.SSSZ. For example, 2019-11-08T02:41:28.172Z.

item_attachments :: Lens' Item (Maybe [AttachmentItem])

Provides information about the attachments.

item_participantRole :: Lens' Item (Maybe ParticipantRole)

The role of the sender. For example, is it a customer, agent, or system.

item_content :: Lens' Item (Maybe Text)

The content of the message or event.

item_id :: Lens' Item (Maybe Text)

The ID of the item.

item_displayName :: Lens' Item (Maybe Text)

The chat display name of the sender.

item_type :: Lens' Item (Maybe ChatItemType)

Type of the item: message or event.

item_contentType :: Lens' Item (Maybe Text)

The type of content of the item.

StartPosition

startPosition_absoluteTime :: Lens' StartPosition (Maybe Text)

The time in ISO format where to start.

It's specified in ISO 8601 format: yyyy-MM-ddThh:mm:ss.SSSZ. For example, 2019-11-08T02:41:28.172Z.

startPosition_id :: Lens' StartPosition (Maybe Text)

The ID of the message or event where to start.

startPosition_mostRecent :: Lens' StartPosition (Maybe Natural)

The start position of the most recent message where you want to start.

UploadMetadata

uploadMetadata_urlExpiry :: Lens' UploadMetadata (Maybe Text)

The expiration time of the URL in ISO timestamp. It's specified in ISO 8601 format: yyyy-MM-ddThh:mm:ss.SSSZ. For example, 2019-11-08T02:41:28.172Z.

uploadMetadata_headersToInclude :: Lens' UploadMetadata (Maybe (HashMap Text Text))

The headers to be provided while uploading the file to the URL.

uploadMetadata_url :: Lens' UploadMetadata (Maybe Text)

The pre-signed URL using which file would be downloaded from Amazon S3 by the API caller.

Websocket

websocket_url :: Lens' Websocket (Maybe Text)

The URL of the websocket.

websocket_connectionExpiry :: Lens' Websocket (Maybe Text)

The URL expiration timestamp in ISO date format.

It's specified in ISO 8601 format: yyyy-MM-ddThh:mm:ss.SSSZ. For example, 2019-11-08T02:41:28.172Z.