Amazonka.ConnectParticipant.Types

Description

Service Configuration

defaultService :: Service

API version 2018-09-07 of the Amazon Connect Participant Service SDK configuration.

Errors

_ValidationException :: AsError a => Getting (First ServiceError) a ServiceError

The input fails to satisfy the constraints specified by Amazon Connect.

_AccessDeniedException :: AsError a => Getting (First ServiceError) a ServiceError

You do not have sufficient access to perform this action.

_ConflictException :: AsError a => Getting (First ServiceError) a ServiceError

An attachment with that identifier is already being uploaded.

_ServiceQuotaExceededException :: AsError a => Getting (First ServiceError) a ServiceError

The number of attachments per contact exceeds the quota.

_ThrottlingException :: AsError a => Getting (First ServiceError) a ServiceError

The request was denied due to request throttling.

_InternalServerException :: AsError a => Getting (First ServiceError) a ServiceError

This exception occurs when there is an internal failure in the Amazon Connect service.

ArtifactStatus

newtype ArtifactStatus

Constructors

ArtifactStatus'

Fields

fromArtifactStatus :: Text

Bundled Patterns

pattern ArtifactStatus_APPROVED :: ArtifactStatus

pattern ArtifactStatus_IN_PROGRESS :: ArtifactStatus

pattern ArtifactStatus_REJECTED :: ArtifactStatus

ChatItemType

newtype ChatItemType

Constructors

ChatItemType'

Fields

fromChatItemType :: Text

Bundled Patterns

pattern ChatItemType_ATTACHMENT :: ChatItemType

pattern ChatItemType_CHAT_ENDED :: ChatItemType

pattern ChatItemType_CONNECTION_ACK :: ChatItemType

pattern ChatItemType_EVENT :: ChatItemType

pattern ChatItemType_MESSAGE :: ChatItemType

pattern ChatItemType_PARTICIPANT_JOINED :: ChatItemType

pattern ChatItemType_PARTICIPANT_LEFT :: ChatItemType

pattern ChatItemType_TRANSFER_FAILED :: ChatItemType

pattern ChatItemType_TRANSFER_SUCCEEDED :: ChatItemType

pattern ChatItemType_TYPING :: ChatItemType

ConnectionType

newtype ConnectionType

Constructors

ConnectionType'

Fields

fromConnectionType :: Text

Bundled Patterns

pattern ConnectionType_CONNECTION_CREDENTIALS :: ConnectionType

pattern ConnectionType_WEBSOCKET :: ConnectionType

ParticipantRole

newtype ParticipantRole

Constructors

ParticipantRole'

Fields

fromParticipantRole :: Text

Bundled Patterns

pattern ParticipantRole_AGENT :: ParticipantRole

pattern ParticipantRole_CUSTOMER :: ParticipantRole

pattern ParticipantRole_SYSTEM :: ParticipantRole

ScanDirection

newtype ScanDirection

Constructors

ScanDirection'

Fields

fromScanDirection :: Text

Bundled Patterns

pattern ScanDirection_BACKWARD :: ScanDirection

pattern ScanDirection_FORWARD :: ScanDirection

SortKey

newtype SortKey

Constructors

SortKey'

Fields

fromSortKey :: Text

Bundled Patterns

pattern SortKey_ASCENDING :: SortKey

pattern SortKey_DESCENDING :: SortKey

AttachmentItem

data AttachmentItem

The case-insensitive input to indicate standard MIME type that describes the format of the file that will be uploaded.

See: newAttachmentItem smart constructor.

Constructors

AttachmentItem'

Fields

status :: Maybe ArtifactStatus

Status of the attachment.

attachmentName :: Maybe Text

A case-sensitive name of the attachment being uploaded.

attachmentId :: Maybe Text

A unique identifier for the attachment.

contentType :: Maybe Text

Describes the MIME file type of the attachment. For a list of supported file types, see Feature specifications in the Amazon Connect Administrator Guide.

newAttachmentItem :: AttachmentItem

Create a value of AttachmentItem with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:status:AttachmentItem', attachmentItem_status - Status of the attachment.

$sel:attachmentName:AttachmentItem', attachmentItem_attachmentName - A case-sensitive name of the attachment being uploaded.

$sel:attachmentId:AttachmentItem', attachmentItem_attachmentId - A unique identifier for the attachment.

$sel:contentType:AttachmentItem', attachmentItem_contentType - Describes the MIME file type of the attachment. For a list of supported file types, see Feature specifications in the Amazon Connect Administrator Guide.

attachmentItem_status :: Lens' AttachmentItem (Maybe ArtifactStatus)

Status of the attachment.

attachmentItem_attachmentName :: Lens' AttachmentItem (Maybe Text)

A case-sensitive name of the attachment being uploaded.

attachmentItem_attachmentId :: Lens' AttachmentItem (Maybe Text)

A unique identifier for the attachment.

attachmentItem_contentType :: Lens' AttachmentItem (Maybe Text)

Describes the MIME file type of the attachment. For a list of supported file types, see Feature specifications in the Amazon Connect Administrator Guide.

ConnectionCredentials

data ConnectionCredentials

Connection credentials.

See: newConnectionCredentials smart constructor.

Constructors

ConnectionCredentials'

Fields

expiry :: Maybe Text

The expiration of the token.

It's specified in ISO 8601 format: yyyy-MM-ddThh:mm:ss.SSSZ. For example, 2019-11-08T02:41:28.172Z.

connectionToken :: Maybe Text

The connection token.

newConnectionCredentials :: ConnectionCredentials

Create a value of ConnectionCredentials with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:expiry:ConnectionCredentials', connectionCredentials_expiry - The expiration of the token.

It's specified in ISO 8601 format: yyyy-MM-ddThh:mm:ss.SSSZ. For example, 2019-11-08T02:41:28.172Z.

$sel:connectionToken:ConnectionCredentials', connectionCredentials_connectionToken - The connection token.

connectionCredentials_expiry :: Lens' ConnectionCredentials (Maybe Text)

The expiration of the token.

It's specified in ISO 8601 format: yyyy-MM-ddThh:mm:ss.SSSZ. For example, 2019-11-08T02:41:28.172Z.

connectionCredentials_connectionToken :: Lens' ConnectionCredentials (Maybe Text)

The connection token.

Item

data Item

An item - message or event - that has been sent.

See: newItem smart constructor.

Constructors

Item'

Fields

participantId :: Maybe Text

The ID of the sender in the session.

absoluteTime :: Maybe Text

The time when the message or event was sent.

It's specified in ISO 8601 format: yyyy-MM-ddThh:mm:ss.SSSZ. For example, 2019-11-08T02:41:28.172Z.

attachments :: Maybe [AttachmentItem]

Provides information about the attachments.

participantRole :: Maybe ParticipantRole

The role of the sender. For example, is it a customer, agent, or system.

content :: Maybe Text

The content of the message or event.

id :: Maybe Text

The ID of the item.

displayName :: Maybe Text

The chat display name of the sender.

type' :: Maybe ChatItemType

Type of the item: message or event.

contentType :: Maybe Text

The type of content of the item.

newItem :: Item

Create a value of Item with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:participantId:Item', item_participantId - The ID of the sender in the session.

$sel:absoluteTime:Item', item_absoluteTime - The time when the message or event was sent.

It's specified in ISO 8601 format: yyyy-MM-ddThh:mm:ss.SSSZ. For example, 2019-11-08T02:41:28.172Z.

$sel:attachments:Item', item_attachments - Provides information about the attachments.

$sel:participantRole:Item', item_participantRole - The role of the sender. For example, is it a customer, agent, or system.

$sel:content:Item', item_content - The content of the message or event.

$sel:id:Item', item_id - The ID of the item.

$sel:displayName:Item', item_displayName - The chat display name of the sender.

$sel:type':Item', item_type - Type of the item: message or event.

$sel:contentType:Item', item_contentType - The type of content of the item.

item_participantId :: Lens' Item (Maybe Text)

The ID of the sender in the session.

item_absoluteTime :: Lens' Item (Maybe Text)

The time when the message or event was sent.

It's specified in ISO 8601 format: yyyy-MM-ddThh:mm:ss.SSSZ. For example, 2019-11-08T02:41:28.172Z.

item_attachments :: Lens' Item (Maybe [AttachmentItem])

Provides information about the attachments.

item_participantRole :: Lens' Item (Maybe ParticipantRole)

The role of the sender. For example, is it a customer, agent, or system.

item_content :: Lens' Item (Maybe Text)

The content of the message or event.

item_id :: Lens' Item (Maybe Text)

The ID of the item.

item_displayName :: Lens' Item (Maybe Text)

The chat display name of the sender.

item_type :: Lens' Item (Maybe ChatItemType)

Type of the item: message or event.

item_contentType :: Lens' Item (Maybe Text)

The type of content of the item.

StartPosition

data StartPosition

A filtering option for where to start. For example, if you sent 100 messages, start with message 50.

See: newStartPosition smart constructor.

Constructors

StartPosition'

Fields

absoluteTime :: Maybe Text

The time in ISO format where to start.

It's specified in ISO 8601 format: yyyy-MM-ddThh:mm:ss.SSSZ. For example, 2019-11-08T02:41:28.172Z.

id :: Maybe Text

The ID of the message or event where to start.

mostRecent :: Maybe Natural

The start position of the most recent message where you want to start.

newStartPosition :: StartPosition

Create a value of StartPosition with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:absoluteTime:StartPosition', startPosition_absoluteTime - The time in ISO format where to start.

It's specified in ISO 8601 format: yyyy-MM-ddThh:mm:ss.SSSZ. For example, 2019-11-08T02:41:28.172Z.

$sel:id:StartPosition', startPosition_id - The ID of the message or event where to start.

$sel:mostRecent:StartPosition', startPosition_mostRecent - The start position of the most recent message where you want to start.

startPosition_absoluteTime :: Lens' StartPosition (Maybe Text)

The time in ISO format where to start.

It's specified in ISO 8601 format: yyyy-MM-ddThh:mm:ss.SSSZ. For example, 2019-11-08T02:41:28.172Z.

startPosition_id :: Lens' StartPosition (Maybe Text)

The ID of the message or event where to start.

startPosition_mostRecent :: Lens' StartPosition (Maybe Natural)

The start position of the most recent message where you want to start.

UploadMetadata

data UploadMetadata

Fields to be used while uploading the attachment.

See: newUploadMetadata smart constructor.

Constructors

UploadMetadata'

Fields

urlExpiry :: Maybe Text

The expiration time of the URL in ISO timestamp. It's specified in ISO 8601 format: yyyy-MM-ddThh:mm:ss.SSSZ. For example, 2019-11-08T02:41:28.172Z.

headersToInclude :: Maybe (HashMap Text Text)

The headers to be provided while uploading the file to the URL.

url :: Maybe Text

The pre-signed URL using which file would be downloaded from Amazon S3 by the API caller.

newUploadMetadata :: UploadMetadata

Create a value of UploadMetadata with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:urlExpiry:UploadMetadata', uploadMetadata_urlExpiry - The expiration time of the URL in ISO timestamp. It's specified in ISO 8601 format: yyyy-MM-ddThh:mm:ss.SSSZ. For example, 2019-11-08T02:41:28.172Z.

$sel:headersToInclude:UploadMetadata', uploadMetadata_headersToInclude - The headers to be provided while uploading the file to the URL.

$sel:url:UploadMetadata', uploadMetadata_url - The pre-signed URL using which file would be downloaded from Amazon S3 by the API caller.

uploadMetadata_urlExpiry :: Lens' UploadMetadata (Maybe Text)

The expiration time of the URL in ISO timestamp. It's specified in ISO 8601 format: yyyy-MM-ddThh:mm:ss.SSSZ. For example, 2019-11-08T02:41:28.172Z.

uploadMetadata_headersToInclude :: Lens' UploadMetadata (Maybe (HashMap Text Text))

The headers to be provided while uploading the file to the URL.

uploadMetadata_url :: Lens' UploadMetadata (Maybe Text)

The pre-signed URL using which file would be downloaded from Amazon S3 by the API caller.

Websocket

data Websocket

The websocket for the participant's connection.

See: newWebsocket smart constructor.

Constructors

Websocket'

Fields

url :: Maybe Text

The URL of the websocket.

connectionExpiry :: Maybe Text

The URL expiration timestamp in ISO date format.

It's specified in ISO 8601 format: yyyy-MM-ddThh:mm:ss.SSSZ. For example, 2019-11-08T02:41:28.172Z.

newWebsocket :: Websocket

Create a value of Websocket with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:url:Websocket', websocket_url - The URL of the websocket.

$sel:connectionExpiry:Websocket', websocket_connectionExpiry - The URL expiration timestamp in ISO date format.

It's specified in ISO 8601 format: yyyy-MM-ddThh:mm:ss.SSSZ. For example, 2019-11-08T02:41:28.172Z.

websocket_url :: Lens' Websocket (Maybe Text)

The URL of the websocket.

websocket_connectionExpiry :: Lens' Websocket (Maybe Text)

The URL expiration timestamp in ISO date format.

It's specified in ISO 8601 format: yyyy-MM-ddThh:mm:ss.SSSZ. For example, 2019-11-08T02:41:28.172Z.