Amazonka.ConnectParticipant.Types.AttachmentItem

Description

Documentation

data AttachmentItem

The case-insensitive input to indicate standard MIME type that describes the format of the file that will be uploaded.

See: newAttachmentItem smart constructor.

Constructors

AttachmentItem'

Fields

status :: Maybe ArtifactStatus

Status of the attachment.

attachmentName :: Maybe Text

A case-sensitive name of the attachment being uploaded.

attachmentId :: Maybe Text

A unique identifier for the attachment.

contentType :: Maybe Text

Describes the MIME file type of the attachment. For a list of supported file types, see Feature specifications in the Amazon Connect Administrator Guide.

newAttachmentItem :: AttachmentItem

Create a value of AttachmentItem with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:status:AttachmentItem', attachmentItem_status - Status of the attachment.

$sel:attachmentName:AttachmentItem', attachmentItem_attachmentName - A case-sensitive name of the attachment being uploaded.

$sel:attachmentId:AttachmentItem', attachmentItem_attachmentId - A unique identifier for the attachment.

$sel:contentType:AttachmentItem', attachmentItem_contentType - Describes the MIME file type of the attachment. For a list of supported file types, see Feature specifications in the Amazon Connect Administrator Guide.

attachmentItem_status :: Lens' AttachmentItem (Maybe ArtifactStatus)

Status of the attachment.

attachmentItem_attachmentName :: Lens' AttachmentItem (Maybe Text)

A case-sensitive name of the attachment being uploaded.

attachmentItem_attachmentId :: Lens' AttachmentItem (Maybe Text)

A unique identifier for the attachment.

attachmentItem_contentType :: Lens' AttachmentItem (Maybe Text)

Describes the MIME file type of the attachment. For a list of supported file types, see Feature specifications in the Amazon Connect Administrator Guide.