Amazonka.ConnectParticipant.Types.Item

Description

Documentation

data Item

An item - message or event - that has been sent.

See: newItem smart constructor.

Constructors

Item'

Fields

participantId :: Maybe Text

The ID of the sender in the session.

absoluteTime :: Maybe Text

The time when the message or event was sent.

It's specified in ISO 8601 format: yyyy-MM-ddThh:mm:ss.SSSZ. For example, 2019-11-08T02:41:28.172Z.

attachments :: Maybe [AttachmentItem]

Provides information about the attachments.

participantRole :: Maybe ParticipantRole

The role of the sender. For example, is it a customer, agent, or system.

content :: Maybe Text

The content of the message or event.

id :: Maybe Text

The ID of the item.

displayName :: Maybe Text

The chat display name of the sender.

type' :: Maybe ChatItemType

Type of the item: message or event.

contentType :: Maybe Text

The type of content of the item.

newItem :: Item

Create a value of Item with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:participantId:Item', item_participantId - The ID of the sender in the session.

$sel:absoluteTime:Item', item_absoluteTime - The time when the message or event was sent.

It's specified in ISO 8601 format: yyyy-MM-ddThh:mm:ss.SSSZ. For example, 2019-11-08T02:41:28.172Z.

$sel:attachments:Item', item_attachments - Provides information about the attachments.

$sel:participantRole:Item', item_participantRole - The role of the sender. For example, is it a customer, agent, or system.

$sel:content:Item', item_content - The content of the message or event.

$sel:id:Item', item_id - The ID of the item.

$sel:displayName:Item', item_displayName - The chat display name of the sender.

$sel:type':Item', item_type - Type of the item: message or event.

$sel:contentType:Item', item_contentType - The type of content of the item.

item_participantId :: Lens' Item (Maybe Text)

The ID of the sender in the session.

item_absoluteTime :: Lens' Item (Maybe Text)

The time when the message or event was sent.

It's specified in ISO 8601 format: yyyy-MM-ddThh:mm:ss.SSSZ. For example, 2019-11-08T02:41:28.172Z.

item_attachments :: Lens' Item (Maybe [AttachmentItem])

Provides information about the attachments.

item_participantRole :: Lens' Item (Maybe ParticipantRole)

The role of the sender. For example, is it a customer, agent, or system.

item_content :: Lens' Item (Maybe Text)

The content of the message or event.

item_id :: Lens' Item (Maybe Text)

The ID of the item.

item_displayName :: Lens' Item (Maybe Text)

The chat display name of the sender.

item_type :: Lens' Item (Maybe ChatItemType)

Type of the item: message or event.

item_contentType :: Lens' Item (Maybe Text)

The type of content of the item.