Amazonka.CostAndUsageReport

Description

Derived from API version 2017-01-06 of the AWS service descriptions, licensed under Apache 2.0.

The AWS Cost and Usage Report API enables you to programmatically create, query, and delete AWS Cost and Usage report definitions.

AWS Cost and Usage reports track the monthly AWS costs and usage associated with your AWS account. The report contains line items for each unique combination of AWS product, usage type, and operation that your AWS account uses. You can configure the AWS Cost and Usage report to show only the data that you want, using the AWS Cost and Usage API.

Service Endpoint

The AWS Cost and Usage Report API provides the following endpoint:

cur.us-east-1.amazonaws.com

Service Configuration

defaultService :: Service

API version 2017-01-06 of the Amazon Cost and Usage Report Service SDK configuration.

Errors

Error matchers are designed for use with the functions provided by Control.Exception.Lens. This allows catching (and rethrowing) service specific errors returned by CostAndUsageReport.

ValidationException

_ValidationException :: AsError a => Getting (First ServiceError) a ServiceError

The input fails to satisfy the constraints specified by an AWS service.

InternalErrorException

_InternalErrorException :: AsError a => Getting (First ServiceError) a ServiceError

An error on the server occurred during the processing of your request. Try again later.

DuplicateReportNameException

_DuplicateReportNameException :: AsError a => Getting (First ServiceError) a ServiceError

A report with the specified name already exists in the account. Specify a different report name.

ReportLimitReachedException

_ReportLimitReachedException :: AsError a => Getting (First ServiceError) a ServiceError

This account already has five reports defined. To define a new report, you must delete an existing report.

Waiters

Waiters poll by repeatedly sending a request until some remote success condition configured by the Wait specification is fulfilled. The Wait specification determines how many attempts should be made, in addition to delay and retry strategies.

Operations

Some AWS operations return results that are incomplete and require subsequent requests in order to obtain the entire result set. The process of sending subsequent requests to continue where a previous request left off is called pagination. For example, the ListObjects operation of Amazon S3 returns up to 1000 objects at a time, and you must send subsequent requests with the appropriate Marker in order to retrieve the next page of results.

Operations that have an AWSPager instance can transparently perform subsequent requests, correctly setting Markers and other request facets to iterate through the entire result set of a truncated API operation. Operations which support this have an additional note in the documentation.

Many operations have the ability to filter results on the server side. See the individual operation parameters for details.

PutReportDefinition

data PutReportDefinition

Creates a Cost and Usage Report.

See: newPutReportDefinition smart constructor.

Constructors

PutReportDefinition' ReportDefinition

newPutReportDefinition

Arguments

:: ReportDefinition
$sel:reportDefinition:PutReportDefinition'

-> PutReportDefinition

Create a value of PutReportDefinition with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:reportDefinition:PutReportDefinition', putReportDefinition_reportDefinition - Represents the output of the PutReportDefinition operation. The content consists of the detailed metadata and data file information.

data PutReportDefinitionResponse

If the action is successful, the service sends back an HTTP 200 response with an empty HTTP body.

See: newPutReportDefinitionResponse smart constructor.

Constructors

PutReportDefinitionResponse' Int

newPutReportDefinitionResponse

Arguments

:: Int
$sel:httpStatus:PutReportDefinitionResponse'

-> PutReportDefinitionResponse

Create a value of PutReportDefinitionResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:httpStatus:PutReportDefinitionResponse', putReportDefinitionResponse_httpStatus - The response's http status code.

DeleteReportDefinition

data DeleteReportDefinition

Deletes the specified report.

See: newDeleteReportDefinition smart constructor.

Constructors

DeleteReportDefinition' (Maybe Text)

newDeleteReportDefinition :: DeleteReportDefinition

Create a value of DeleteReportDefinition with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:reportName:DeleteReportDefinition', deleteReportDefinition_reportName - The name of the report that you want to delete. The name must be unique, is case sensitive, and can't include spaces.

data DeleteReportDefinitionResponse

If the action is successful, the service sends back an HTTP 200 response.

See: newDeleteReportDefinitionResponse smart constructor.

Constructors

DeleteReportDefinitionResponse' (Maybe Text) Int

newDeleteReportDefinitionResponse

Arguments

:: Int
$sel:httpStatus:DeleteReportDefinitionResponse'

-> DeleteReportDefinitionResponse

Create a value of DeleteReportDefinitionResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:responseMessage:DeleteReportDefinitionResponse', deleteReportDefinitionResponse_responseMessage - Undocumented member.

$sel:httpStatus:DeleteReportDefinitionResponse', deleteReportDefinitionResponse_httpStatus - The response's http status code.

ModifyReportDefinition

data ModifyReportDefinition

See: newModifyReportDefinition smart constructor.

Constructors

ModifyReportDefinition' Text ReportDefinition

newModifyReportDefinition

Arguments

:: Text
$sel:reportName:ModifyReportDefinition'

-> ReportDefinition
$sel:reportDefinition:ModifyReportDefinition'

-> ModifyReportDefinition

Create a value of ModifyReportDefinition with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:reportName:ModifyReportDefinition', modifyReportDefinition_reportName - Undocumented member.

$sel:reportDefinition:ModifyReportDefinition', modifyReportDefinition_reportDefinition - Undocumented member.

data ModifyReportDefinitionResponse

See: newModifyReportDefinitionResponse smart constructor.

Constructors

ModifyReportDefinitionResponse' Int

newModifyReportDefinitionResponse

Arguments

:: Int
$sel:httpStatus:ModifyReportDefinitionResponse'

-> ModifyReportDefinitionResponse

Create a value of ModifyReportDefinitionResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:httpStatus:ModifyReportDefinitionResponse', modifyReportDefinitionResponse_httpStatus - The response's http status code.

DescribeReportDefinitions (Paginated)

data DescribeReportDefinitions

Requests a list of AWS Cost and Usage reports owned by the account.

See: newDescribeReportDefinitions smart constructor.

Constructors

DescribeReportDefinitions' (Maybe Text) (Maybe Natural)

newDescribeReportDefinitions :: DescribeReportDefinitions

Create a value of DescribeReportDefinitions with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:nextToken:DescribeReportDefinitions', describeReportDefinitions_nextToken - Undocumented member.

$sel:maxResults:DescribeReportDefinitions', describeReportDefinitions_maxResults - Undocumented member.

data DescribeReportDefinitionsResponse

If the action is successful, the service sends back an HTTP 200 response.

See: newDescribeReportDefinitionsResponse smart constructor.

Constructors

DescribeReportDefinitionsResponse' (Maybe Text) (Maybe [ReportDefinition]) Int

newDescribeReportDefinitionsResponse

Arguments

:: Int
$sel:httpStatus:DescribeReportDefinitionsResponse'

-> DescribeReportDefinitionsResponse

Create a value of DescribeReportDefinitionsResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:nextToken:DescribeReportDefinitions', describeReportDefinitionsResponse_nextToken - Undocumented member.

$sel:reportDefinitions:DescribeReportDefinitionsResponse', describeReportDefinitionsResponse_reportDefinitions - A list of AWS Cost and Usage reports owned by the account.

$sel:httpStatus:DescribeReportDefinitionsResponse', describeReportDefinitionsResponse_httpStatus - The response's http status code.

Types

AWSRegion

newtype AWSRegion

The region of the S3 bucket that AWS delivers the report into.

Constructors

AWSRegion'

Fields

fromAWSRegion :: Text

Bundled Patterns

pattern AWSRegion_Af_south_1 :: AWSRegion

pattern AWSRegion_Ap_east_1 :: AWSRegion

pattern AWSRegion_Ap_northeast_1 :: AWSRegion

pattern AWSRegion_Ap_northeast_2 :: AWSRegion

pattern AWSRegion_Ap_northeast_3 :: AWSRegion

pattern AWSRegion_Ap_south_1 :: AWSRegion

pattern AWSRegion_Ap_southeast_1 :: AWSRegion

pattern AWSRegion_Ap_southeast_2 :: AWSRegion

pattern AWSRegion_Ca_central_1 :: AWSRegion

pattern AWSRegion_Cn_north_1 :: AWSRegion

pattern AWSRegion_Cn_northwest_1 :: AWSRegion

pattern AWSRegion_Eu_central_1 :: AWSRegion

pattern AWSRegion_Eu_north_1 :: AWSRegion

pattern AWSRegion_Eu_south_1 :: AWSRegion

pattern AWSRegion_Eu_west_1 :: AWSRegion

pattern AWSRegion_Eu_west_2 :: AWSRegion

pattern AWSRegion_Eu_west_3 :: AWSRegion

pattern AWSRegion_Me_south_1 :: AWSRegion

pattern AWSRegion_Sa_east_1 :: AWSRegion

pattern AWSRegion_Us_east_1 :: AWSRegion

pattern AWSRegion_Us_east_2 :: AWSRegion

pattern AWSRegion_Us_west_1 :: AWSRegion

pattern AWSRegion_Us_west_2 :: AWSRegion

AdditionalArtifact

newtype AdditionalArtifact

The types of manifest that you want AWS to create for this report.

Constructors

AdditionalArtifact'

Fields

fromAdditionalArtifact :: Text

Bundled Patterns

pattern AdditionalArtifact_ATHENA :: AdditionalArtifact

pattern AdditionalArtifact_QUICKSIGHT :: AdditionalArtifact

pattern AdditionalArtifact_REDSHIFT :: AdditionalArtifact

CompressionFormat

newtype CompressionFormat

The compression format that AWS uses for the report.

Constructors

CompressionFormat'

Fields

fromCompressionFormat :: Text

Bundled Patterns

pattern CompressionFormat_GZIP :: CompressionFormat

pattern CompressionFormat_Parquet :: CompressionFormat

pattern CompressionFormat_ZIP :: CompressionFormat

ReportFormat

newtype ReportFormat

The format that AWS saves the report in.

Constructors

ReportFormat'

Fields

fromReportFormat :: Text

Bundled Patterns

pattern ReportFormat_Parquet :: ReportFormat

pattern ReportFormat_TextORcsv :: ReportFormat

ReportVersioning

newtype ReportVersioning

Constructors

ReportVersioning'

Fields

fromReportVersioning :: Text

Bundled Patterns

pattern ReportVersioning_CREATE_NEW_REPORT :: ReportVersioning

pattern ReportVersioning_OVERWRITE_REPORT :: ReportVersioning

SchemaElement

newtype SchemaElement

Whether or not AWS includes resource IDs in the report.

Constructors

SchemaElement'

Fields

fromSchemaElement :: Text

Bundled Patterns

pattern SchemaElement_RESOURCES :: SchemaElement

TimeUnit

newtype TimeUnit

The length of time covered by the report.

Constructors

TimeUnit'

Fields

fromTimeUnit :: Text

Bundled Patterns

pattern TimeUnit_DAILY :: TimeUnit

pattern TimeUnit_HOURLY :: TimeUnit

pattern TimeUnit_MONTHLY :: TimeUnit

ReportDefinition

data ReportDefinition

The definition of AWS Cost and Usage Report. You can specify the report name, time unit, report format, compression format, S3 bucket, additional artifacts, and schema elements in the definition.

See: newReportDefinition smart constructor.

Constructors

ReportDefinition' (Maybe ReportVersioning) (Maybe Text) (Maybe [AdditionalArtifact]) (Maybe Bool) Text TimeUnit ReportFormat CompressionFormat [SchemaElement] Text Text AWSRegion

newReportDefinition

Arguments

:: Text
$sel:reportName:ReportDefinition'

-> TimeUnit
$sel:timeUnit:ReportDefinition'

-> ReportFormat
$sel:format:ReportDefinition'

-> CompressionFormat
$sel:compression:ReportDefinition'

-> Text
$sel:s3Bucket:ReportDefinition'

-> Text
$sel:s3Prefix:ReportDefinition'

-> AWSRegion
$sel:s3Region:ReportDefinition'

-> ReportDefinition

Create a value of ReportDefinition with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:reportVersioning:ReportDefinition', reportDefinition_reportVersioning - Whether you want Amazon Web Services to overwrite the previous version of each report or to deliver the report in addition to the previous versions.

$sel:billingViewArn:ReportDefinition', reportDefinition_billingViewArn - The Amazon resource name of the billing view. You can get this value by using the billing view service public APIs.

$sel:additionalArtifacts:ReportDefinition', reportDefinition_additionalArtifacts - A list of manifests that you want Amazon Web Services to create for this report.

$sel:refreshClosedReports:ReportDefinition', reportDefinition_refreshClosedReports - Whether you want Amazon Web Services to update your reports after they have been finalized if Amazon Web Services detects charges related to previous months. These charges can include refunds, credits, or support fees.

$sel:reportName:ReportDefinition', reportDefinition_reportName - Undocumented member.

$sel:timeUnit:ReportDefinition', reportDefinition_timeUnit - Undocumented member.

$sel:format:ReportDefinition', reportDefinition_format - Undocumented member.

$sel:compression:ReportDefinition', reportDefinition_compression - Undocumented member.

$sel:additionalSchemaElements:ReportDefinition', reportDefinition_additionalSchemaElements - A list of strings that indicate additional content that Amazon Web Services includes in the report, such as individual resource IDs.

$sel:s3Bucket:ReportDefinition', reportDefinition_s3Bucket - Undocumented member.

$sel:s3Prefix:ReportDefinition', reportDefinition_s3Prefix - Undocumented member.

$sel:s3Region:ReportDefinition', reportDefinition_s3Region - Undocumented member.