Amazonka.CostAndUsageReport.Lens

Description

Operations

PutReportDefinition

putReportDefinition_reportDefinition :: Lens' PutReportDefinition ReportDefinition

Represents the output of the PutReportDefinition operation. The content consists of the detailed metadata and data file information.

putReportDefinitionResponse_httpStatus :: Lens' PutReportDefinitionResponse Int

The response's http status code.

DeleteReportDefinition

deleteReportDefinition_reportName :: Lens' DeleteReportDefinition (Maybe Text)

The name of the report that you want to delete. The name must be unique, is case sensitive, and can't include spaces.

deleteReportDefinitionResponse_responseMessage :: Lens' DeleteReportDefinitionResponse (Maybe Text)

Undocumented member.

deleteReportDefinitionResponse_httpStatus :: Lens' DeleteReportDefinitionResponse Int

The response's http status code.

ModifyReportDefinition

modifyReportDefinition_reportName :: Lens' ModifyReportDefinition Text

Undocumented member.

modifyReportDefinition_reportDefinition :: Lens' ModifyReportDefinition ReportDefinition

Undocumented member.

modifyReportDefinitionResponse_httpStatus :: Lens' ModifyReportDefinitionResponse Int

The response's http status code.

DescribeReportDefinitions

describeReportDefinitions_nextToken :: Lens' DescribeReportDefinitions (Maybe Text)

Undocumented member.

describeReportDefinitions_maxResults :: Lens' DescribeReportDefinitions (Maybe Natural)

Undocumented member.

describeReportDefinitionsResponse_nextToken :: Lens' DescribeReportDefinitionsResponse (Maybe Text)

Undocumented member.

describeReportDefinitionsResponse_reportDefinitions :: Lens' DescribeReportDefinitionsResponse (Maybe [ReportDefinition])

A list of AWS Cost and Usage reports owned by the account.

describeReportDefinitionsResponse_httpStatus :: Lens' DescribeReportDefinitionsResponse Int

The response's http status code.

Types

ReportDefinition

reportDefinition_reportVersioning :: Lens' ReportDefinition (Maybe ReportVersioning)

Whether you want Amazon Web Services to overwrite the previous version of each report or to deliver the report in addition to the previous versions.

reportDefinition_billingViewArn :: Lens' ReportDefinition (Maybe Text)

The Amazon resource name of the billing view. You can get this value by using the billing view service public APIs.

reportDefinition_additionalArtifacts :: Lens' ReportDefinition (Maybe [AdditionalArtifact])

A list of manifests that you want Amazon Web Services to create for this report.

reportDefinition_refreshClosedReports :: Lens' ReportDefinition (Maybe Bool)

Whether you want Amazon Web Services to update your reports after they have been finalized if Amazon Web Services detects charges related to previous months. These charges can include refunds, credits, or support fees.

reportDefinition_reportName :: Lens' ReportDefinition Text

Undocumented member.

reportDefinition_timeUnit :: Lens' ReportDefinition TimeUnit

Undocumented member.

reportDefinition_format :: Lens' ReportDefinition ReportFormat

Undocumented member.

reportDefinition_compression :: Lens' ReportDefinition CompressionFormat

Undocumented member.

reportDefinition_additionalSchemaElements :: Lens' ReportDefinition [SchemaElement]

A list of strings that indicate additional content that Amazon Web Services includes in the report, such as individual resource IDs.

reportDefinition_s3Bucket :: Lens' ReportDefinition Text

Undocumented member.

reportDefinition_s3Prefix :: Lens' ReportDefinition Text

Undocumented member.

reportDefinition_s3Region :: Lens' ReportDefinition AWSRegion

Undocumented member.