Amazonka.CostExplorer.GetAnomalySubscriptions

Description

Retrieves the cost anomaly subscription objects for your account. You can filter using a list of cost anomaly monitor Amazon Resource Names (ARNs).

Creating a Request

data GetAnomalySubscriptions

See: newGetAnomalySubscriptions smart constructor.

Constructors

GetAnomalySubscriptions'

Fields

subscriptionArnList :: Maybe [Text]

A list of cost anomaly subscription ARNs.

nextPageToken :: Maybe Text

The token to retrieve the next set of results. Amazon Web Services provides the token when the response from a previous call has more results than the maximum page size.

maxResults :: Maybe Int

The number of entries a paginated response contains.

monitorArn :: Maybe Text

Cost anomaly monitor ARNs.

newGetAnomalySubscriptions :: GetAnomalySubscriptions

Create a value of GetAnomalySubscriptions with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:subscriptionArnList:GetAnomalySubscriptions', getAnomalySubscriptions_subscriptionArnList - A list of cost anomaly subscription ARNs.

$sel:nextPageToken:GetAnomalySubscriptions', getAnomalySubscriptions_nextPageToken - The token to retrieve the next set of results. Amazon Web Services provides the token when the response from a previous call has more results than the maximum page size.

$sel:maxResults:GetAnomalySubscriptions', getAnomalySubscriptions_maxResults - The number of entries a paginated response contains.

$sel:monitorArn:GetAnomalySubscriptions', getAnomalySubscriptions_monitorArn - Cost anomaly monitor ARNs.

Request Lenses

getAnomalySubscriptions_subscriptionArnList :: Lens' GetAnomalySubscriptions (Maybe [Text])

A list of cost anomaly subscription ARNs.

getAnomalySubscriptions_nextPageToken :: Lens' GetAnomalySubscriptions (Maybe Text)

The token to retrieve the next set of results. Amazon Web Services provides the token when the response from a previous call has more results than the maximum page size.

getAnomalySubscriptions_maxResults :: Lens' GetAnomalySubscriptions (Maybe Int)

The number of entries a paginated response contains.

getAnomalySubscriptions_monitorArn :: Lens' GetAnomalySubscriptions (Maybe Text)

Cost anomaly monitor ARNs.

Destructuring the Response

data GetAnomalySubscriptionsResponse

See: newGetAnomalySubscriptionsResponse smart constructor.

Constructors

GetAnomalySubscriptionsResponse'

Fields

nextPageToken :: Maybe Text

The token to retrieve the next set of results. Amazon Web Services provides the token when the response from a previous call has more results than the maximum page size.

httpStatus :: Int

The response's http status code.

anomalySubscriptions :: [AnomalySubscription]

A list of cost anomaly subscriptions that includes the detailed metadata for each one.

newGetAnomalySubscriptionsResponse

Arguments

:: Int
$sel:httpStatus:GetAnomalySubscriptionsResponse'

-> GetAnomalySubscriptionsResponse

Create a value of GetAnomalySubscriptionsResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:nextPageToken:GetAnomalySubscriptions', getAnomalySubscriptionsResponse_nextPageToken - The token to retrieve the next set of results. Amazon Web Services provides the token when the response from a previous call has more results than the maximum page size.

$sel:httpStatus:GetAnomalySubscriptionsResponse', getAnomalySubscriptionsResponse_httpStatus - The response's http status code.

$sel:anomalySubscriptions:GetAnomalySubscriptionsResponse', getAnomalySubscriptionsResponse_anomalySubscriptions - A list of cost anomaly subscriptions that includes the detailed metadata for each one.

Response Lenses

getAnomalySubscriptionsResponse_nextPageToken :: Lens' GetAnomalySubscriptionsResponse (Maybe Text)

The token to retrieve the next set of results. Amazon Web Services provides the token when the response from a previous call has more results than the maximum page size.

getAnomalySubscriptionsResponse_httpStatus :: Lens' GetAnomalySubscriptionsResponse Int

The response's http status code.

getAnomalySubscriptionsResponse_anomalySubscriptions :: Lens' GetAnomalySubscriptionsResponse [AnomalySubscription]

A list of cost anomaly subscriptions that includes the detailed metadata for each one.