Amazonka.CostExplorer.GetReservationPurchaseRecommendation

Description

Gets recommendations for which reservations to purchase. These recommendations could help you reduce your costs. Reservations provide a discounted hourly rate (up to 75%) compared to On-Demand pricing.

Amazon Web Services generates your recommendations by identifying your On-Demand usage during a specific time period and collecting your usage into categories that are eligible for a reservation. After Amazon Web Services has these categories, it simulates every combination of reservations in each category of usage to identify the best number of each type of RI to purchase to maximize your estimated savings.

For example, Amazon Web Services automatically aggregates your Amazon EC2 Linux, shared tenancy, and c4 family usage in the US West (Oregon) Region and recommends that you buy size-flexible regional reservations to apply to the c4 family usage. Amazon Web Services recommends the smallest size instance in an instance family. This makes it easier to purchase a size-flexible RI. Amazon Web Services also shows the equal number of normalized units so that you can purchase any instance size that you want. For this example, your RI recommendation would be for c4.large because that is the smallest size instance in the c4 instance family.

Creating a Request

data GetReservationPurchaseRecommendation

See: newGetReservationPurchaseRecommendation smart constructor.

Constructors

GetReservationPurchaseRecommendation'

Fields

nextPageToken :: Maybe Text

The pagination token that indicates the next set of results that you want to retrieve.

termInYears :: Maybe TermInYears

The reservation term that you want recommendations for.

serviceSpecification :: Maybe ServiceSpecification

The hardware specifications for the service instances that you want recommendations for, such as standard or convertible Amazon EC2 instances.

accountScope :: Maybe AccountScope

The account scope that you want your recommendations for. Amazon Web Services calculates recommendations including the management account and member accounts if the value is set to PAYER. If the value is LINKED, recommendations are calculated for individual member accounts only.

accountId :: Maybe Text

The account ID that is associated with the recommendation.

filter' :: Maybe Expression

pageSize :: Maybe Natural

The number of recommendations that you want returned in a single response object.

lookbackPeriodInDays :: Maybe LookbackPeriodInDays

The number of previous days that you want Amazon Web Services to consider when it calculates your recommendations.

paymentOption :: Maybe PaymentOption

The reservation purchase option that you want recommendations for.

service :: Text

The specific service that you want recommendations for.

newGetReservationPurchaseRecommendation

Arguments

:: Text
$sel:service:GetReservationPurchaseRecommendation'

-> GetReservationPurchaseRecommendation

Create a value of GetReservationPurchaseRecommendation with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:nextPageToken:GetReservationPurchaseRecommendation', getReservationPurchaseRecommendation_nextPageToken - The pagination token that indicates the next set of results that you want to retrieve.

$sel:termInYears:GetReservationPurchaseRecommendation', getReservationPurchaseRecommendation_termInYears - The reservation term that you want recommendations for.

$sel:serviceSpecification:GetReservationPurchaseRecommendation', getReservationPurchaseRecommendation_serviceSpecification - The hardware specifications for the service instances that you want recommendations for, such as standard or convertible Amazon EC2 instances.

$sel:accountScope:GetReservationPurchaseRecommendation', getReservationPurchaseRecommendation_accountScope - The account scope that you want your recommendations for. Amazon Web Services calculates recommendations including the management account and member accounts if the value is set to PAYER. If the value is LINKED, recommendations are calculated for individual member accounts only.

$sel:accountId:GetReservationPurchaseRecommendation', getReservationPurchaseRecommendation_accountId - The account ID that is associated with the recommendation.

$sel:filter':GetReservationPurchaseRecommendation', getReservationPurchaseRecommendation_filter - Undocumented member.

$sel:pageSize:GetReservationPurchaseRecommendation', getReservationPurchaseRecommendation_pageSize - The number of recommendations that you want returned in a single response object.

$sel:lookbackPeriodInDays:GetReservationPurchaseRecommendation', getReservationPurchaseRecommendation_lookbackPeriodInDays - The number of previous days that you want Amazon Web Services to consider when it calculates your recommendations.

$sel:paymentOption:GetReservationPurchaseRecommendation', getReservationPurchaseRecommendation_paymentOption - The reservation purchase option that you want recommendations for.

$sel:service:GetReservationPurchaseRecommendation', getReservationPurchaseRecommendation_service - The specific service that you want recommendations for.

Request Lenses

getReservationPurchaseRecommendation_nextPageToken :: Lens' GetReservationPurchaseRecommendation (Maybe Text)

The pagination token that indicates the next set of results that you want to retrieve.

getReservationPurchaseRecommendation_termInYears :: Lens' GetReservationPurchaseRecommendation (Maybe TermInYears)

The reservation term that you want recommendations for.

getReservationPurchaseRecommendation_serviceSpecification :: Lens' GetReservationPurchaseRecommendation (Maybe ServiceSpecification)

The hardware specifications for the service instances that you want recommendations for, such as standard or convertible Amazon EC2 instances.

getReservationPurchaseRecommendation_accountScope :: Lens' GetReservationPurchaseRecommendation (Maybe AccountScope)

The account scope that you want your recommendations for. Amazon Web Services calculates recommendations including the management account and member accounts if the value is set to PAYER. If the value is LINKED, recommendations are calculated for individual member accounts only.

getReservationPurchaseRecommendation_accountId :: Lens' GetReservationPurchaseRecommendation (Maybe Text)

The account ID that is associated with the recommendation.

getReservationPurchaseRecommendation_filter :: Lens' GetReservationPurchaseRecommendation (Maybe Expression)

Undocumented member.

getReservationPurchaseRecommendation_pageSize :: Lens' GetReservationPurchaseRecommendation (Maybe Natural)

The number of recommendations that you want returned in a single response object.

getReservationPurchaseRecommendation_lookbackPeriodInDays :: Lens' GetReservationPurchaseRecommendation (Maybe LookbackPeriodInDays)

The number of previous days that you want Amazon Web Services to consider when it calculates your recommendations.

getReservationPurchaseRecommendation_paymentOption :: Lens' GetReservationPurchaseRecommendation (Maybe PaymentOption)

The reservation purchase option that you want recommendations for.

getReservationPurchaseRecommendation_service :: Lens' GetReservationPurchaseRecommendation Text

The specific service that you want recommendations for.

Destructuring the Response

data GetReservationPurchaseRecommendationResponse

See: newGetReservationPurchaseRecommendationResponse smart constructor.

Constructors

GetReservationPurchaseRecommendationResponse'

Fields

nextPageToken :: Maybe Text

The pagination token for the next set of retrievable results.

recommendations :: Maybe [ReservationPurchaseRecommendation]

Recommendations for reservations to purchase.

metadata :: Maybe ReservationPurchaseRecommendationMetadata

Information about this specific recommendation call, such as the time stamp for when Cost Explorer generated this recommendation.

httpStatus :: Int

The response's http status code.

newGetReservationPurchaseRecommendationResponse

Arguments

:: Int
$sel:httpStatus:GetReservationPurchaseRecommendationResponse'

-> GetReservationPurchaseRecommendationResponse

Create a value of GetReservationPurchaseRecommendationResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:nextPageToken:GetReservationPurchaseRecommendation', getReservationPurchaseRecommendationResponse_nextPageToken - The pagination token for the next set of retrievable results.

$sel:recommendations:GetReservationPurchaseRecommendationResponse', getReservationPurchaseRecommendationResponse_recommendations - Recommendations for reservations to purchase.

$sel:metadata:GetReservationPurchaseRecommendationResponse', getReservationPurchaseRecommendationResponse_metadata - Information about this specific recommendation call, such as the time stamp for when Cost Explorer generated this recommendation.

$sel:httpStatus:GetReservationPurchaseRecommendationResponse', getReservationPurchaseRecommendationResponse_httpStatus - The response's http status code.

Response Lenses

getReservationPurchaseRecommendationResponse_nextPageToken :: Lens' GetReservationPurchaseRecommendationResponse (Maybe Text)

The pagination token for the next set of retrievable results.

getReservationPurchaseRecommendationResponse_recommendations :: Lens' GetReservationPurchaseRecommendationResponse (Maybe [ReservationPurchaseRecommendation])

Recommendations for reservations to purchase.

getReservationPurchaseRecommendationResponse_metadata :: Lens' GetReservationPurchaseRecommendationResponse (Maybe ReservationPurchaseRecommendationMetadata)

Information about this specific recommendation call, such as the time stamp for when Cost Explorer generated this recommendation.

getReservationPurchaseRecommendationResponse_httpStatus :: Lens' GetReservationPurchaseRecommendationResponse Int

The response's http status code.