Amazonka.CostExplorer.GetSavingsPlansUtilization

Description

Retrieves the Savings Plans utilization for your account across date ranges with daily or monthly granularity. Management account in an organization have access to member accounts. You can use GetDimensionValues in SAVINGS_PLANS to determine the possible dimension values.

You cannot group by any dimension values for GetSavingsPlansUtilization.

Creating a Request

data GetSavingsPlansUtilization

See: newGetSavingsPlansUtilization smart constructor.

Constructors

GetSavingsPlansUtilization'

Fields

granularity :: Maybe Granularity

The granularity of the Amazon Web Services utillization data for your Savings Plans.

The GetSavingsPlansUtilization operation supports only DAILY and MONTHLY granularities.

filter' :: Maybe Expression

Filters Savings Plans utilization coverage data for active Savings Plans dimensions. You can filter data with the following dimensions:

LINKED_ACCOUNT

SAVINGS_PLAN_ARN

SAVINGS_PLANS_TYPE

REGION

PAYMENT_OPTION

INSTANCE_TYPE_FAMILY

GetSavingsPlansUtilization uses the same Expression object as the other operations, but only AND is supported among each dimension.

sortBy :: Maybe SortDefinition

The value by which you want to sort the data.

The following values are supported for Key:

UtilizationPercentage

TotalCommitment

UsedCommitment

UnusedCommitment

NetSavings

Supported values for SortOrder are ASCENDING or DESCENDING.

timePeriod :: DateInterval

The time period that you want the usage and costs for. The Start date must be within 13 months. The End date must be after the Start date, and before the current date. Future dates can't be used as an End date.

newGetSavingsPlansUtilization

Arguments

:: DateInterval
$sel:timePeriod:GetSavingsPlansUtilization'

-> GetSavingsPlansUtilization

Create a value of GetSavingsPlansUtilization with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:granularity:GetSavingsPlansUtilization', getSavingsPlansUtilization_granularity - The granularity of the Amazon Web Services utillization data for your Savings Plans.

The GetSavingsPlansUtilization operation supports only DAILY and MONTHLY granularities.

$sel:filter':GetSavingsPlansUtilization', getSavingsPlansUtilization_filter - Filters Savings Plans utilization coverage data for active Savings Plans dimensions. You can filter data with the following dimensions:

LINKED_ACCOUNT

SAVINGS_PLAN_ARN

SAVINGS_PLANS_TYPE

REGION

PAYMENT_OPTION

INSTANCE_TYPE_FAMILY

GetSavingsPlansUtilization uses the same Expression object as the other operations, but only AND is supported among each dimension.

$sel:sortBy:GetSavingsPlansUtilization', getSavingsPlansUtilization_sortBy - The value by which you want to sort the data.

The following values are supported for Key:

UtilizationPercentage

TotalCommitment

UsedCommitment

UnusedCommitment

NetSavings

Supported values for SortOrder are ASCENDING or DESCENDING.

$sel:timePeriod:GetSavingsPlansUtilization', getSavingsPlansUtilization_timePeriod - The time period that you want the usage and costs for. The Start date must be within 13 months. The End date must be after the Start date, and before the current date. Future dates can't be used as an End date.

Request Lenses

getSavingsPlansUtilization_granularity :: Lens' GetSavingsPlansUtilization (Maybe Granularity)

The granularity of the Amazon Web Services utillization data for your Savings Plans.

The GetSavingsPlansUtilization operation supports only DAILY and MONTHLY granularities.

getSavingsPlansUtilization_filter :: Lens' GetSavingsPlansUtilization (Maybe Expression)

Filters Savings Plans utilization coverage data for active Savings Plans dimensions. You can filter data with the following dimensions:

LINKED_ACCOUNT

SAVINGS_PLAN_ARN

SAVINGS_PLANS_TYPE

REGION

PAYMENT_OPTION

INSTANCE_TYPE_FAMILY

GetSavingsPlansUtilization uses the same Expression object as the other operations, but only AND is supported among each dimension.

getSavingsPlansUtilization_sortBy :: Lens' GetSavingsPlansUtilization (Maybe SortDefinition)

The value by which you want to sort the data.

The following values are supported for Key:

UtilizationPercentage

TotalCommitment

UsedCommitment

UnusedCommitment

NetSavings

Supported values for SortOrder are ASCENDING or DESCENDING.

getSavingsPlansUtilization_timePeriod :: Lens' GetSavingsPlansUtilization DateInterval

The time period that you want the usage and costs for. The Start date must be within 13 months. The End date must be after the Start date, and before the current date. Future dates can't be used as an End date.

Destructuring the Response

data GetSavingsPlansUtilizationResponse

See: newGetSavingsPlansUtilizationResponse smart constructor.

Constructors

GetSavingsPlansUtilizationResponse'

Fields

savingsPlansUtilizationsByTime :: Maybe [SavingsPlansUtilizationByTime]

The amount of cost/commitment you used your Savings Plans. This allows you to specify date ranges.

httpStatus :: Int

The response's http status code.

total :: SavingsPlansUtilizationAggregates

The total amount of cost/commitment that you used your Savings Plans, regardless of date ranges.

newGetSavingsPlansUtilizationResponse

Arguments

:: Int
$sel:httpStatus:GetSavingsPlansUtilizationResponse'

-> SavingsPlansUtilizationAggregates
$sel:total:GetSavingsPlansUtilizationResponse'

-> GetSavingsPlansUtilizationResponse

Create a value of GetSavingsPlansUtilizationResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:savingsPlansUtilizationsByTime:GetSavingsPlansUtilizationResponse', getSavingsPlansUtilizationResponse_savingsPlansUtilizationsByTime - The amount of cost/commitment you used your Savings Plans. This allows you to specify date ranges.

$sel:httpStatus:GetSavingsPlansUtilizationResponse', getSavingsPlansUtilizationResponse_httpStatus - The response's http status code.

$sel:total:GetSavingsPlansUtilizationResponse', getSavingsPlansUtilizationResponse_total - The total amount of cost/commitment that you used your Savings Plans, regardless of date ranges.

Response Lenses

getSavingsPlansUtilizationResponse_savingsPlansUtilizationsByTime :: Lens' GetSavingsPlansUtilizationResponse (Maybe [SavingsPlansUtilizationByTime])

The amount of cost/commitment you used your Savings Plans. This allows you to specify date ranges.

getSavingsPlansUtilizationResponse_httpStatus :: Lens' GetSavingsPlansUtilizationResponse Int

The response's http status code.

getSavingsPlansUtilizationResponse_total :: Lens' GetSavingsPlansUtilizationResponse SavingsPlansUtilizationAggregates

The total amount of cost/commitment that you used your Savings Plans, regardless of date ranges.