Amazonka.CostExplorer.GetSavingsPlansUtilizationDetails

Description

Retrieves attribute data along with aggregate utilization and savings data for a given time period. This doesn't support granular or grouped data (daily/monthly) in response. You can't retrieve data by dates in a single response similar to GetSavingsPlanUtilization, but you have the option to make multiple calls to GetSavingsPlanUtilizationDetails by providing individual dates. You can use GetDimensionValues in SAVINGS_PLANS to determine the possible dimension values.

GetSavingsPlanUtilizationDetails internally groups data by SavingsPlansArn.

Creating a Request

data GetSavingsPlansUtilizationDetails

See: newGetSavingsPlansUtilizationDetails smart constructor.

Constructors

GetSavingsPlansUtilizationDetails'

Fields

nextToken :: Maybe Text

The token to retrieve the next set of results. Amazon Web Services provides the token when the response from a previous call has more results than the maximum page size.

dataType :: Maybe [SavingsPlansDataType]

The data type.

filter' :: Maybe Expression

Filters Savings Plans utilization coverage data for active Savings Plans dimensions. You can filter data with the following dimensions:

LINKED_ACCOUNT

SAVINGS_PLAN_ARN

REGION

PAYMENT_OPTION

INSTANCE_TYPE_FAMILY

GetSavingsPlansUtilizationDetails uses the same Expression object as the other operations, but only AND is supported among each dimension.

maxResults :: Maybe Natural

The number of items to be returned in a response. The default is 20, with a minimum value of 1.

sortBy :: Maybe SortDefinition

The value by which you want to sort the data.

The following values are supported for Key:

UtilizationPercentage

TotalCommitment

UsedCommitment

UnusedCommitment

NetSavings

AmortizedRecurringCommitment

AmortizedUpfrontCommitment

Supported values for SortOrder are ASCENDING or DESCENDING.

timePeriod :: DateInterval

The time period that you want the usage and costs for. The Start date must be within 13 months. The End date must be after the Start date, and before the current date. Future dates can't be used as an End date.

newGetSavingsPlansUtilizationDetails

Arguments

:: DateInterval
$sel:timePeriod:GetSavingsPlansUtilizationDetails'

-> GetSavingsPlansUtilizationDetails

Create a value of GetSavingsPlansUtilizationDetails with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:nextToken:GetSavingsPlansUtilizationDetails', getSavingsPlansUtilizationDetails_nextToken - The token to retrieve the next set of results. Amazon Web Services provides the token when the response from a previous call has more results than the maximum page size.

$sel:dataType:GetSavingsPlansUtilizationDetails', getSavingsPlansUtilizationDetails_dataType - The data type.

$sel:filter':GetSavingsPlansUtilizationDetails', getSavingsPlansUtilizationDetails_filter - Filters Savings Plans utilization coverage data for active Savings Plans dimensions. You can filter data with the following dimensions:

LINKED_ACCOUNT

SAVINGS_PLAN_ARN

REGION

PAYMENT_OPTION

INSTANCE_TYPE_FAMILY

GetSavingsPlansUtilizationDetails uses the same Expression object as the other operations, but only AND is supported among each dimension.

$sel:maxResults:GetSavingsPlansUtilizationDetails', getSavingsPlansUtilizationDetails_maxResults - The number of items to be returned in a response. The default is 20, with a minimum value of 1.

$sel:sortBy:GetSavingsPlansUtilizationDetails', getSavingsPlansUtilizationDetails_sortBy - The value by which you want to sort the data.

The following values are supported for Key:

UtilizationPercentage

TotalCommitment

UsedCommitment

UnusedCommitment

NetSavings

AmortizedRecurringCommitment

AmortizedUpfrontCommitment

Supported values for SortOrder are ASCENDING or DESCENDING.

$sel:timePeriod:GetSavingsPlansUtilizationDetails', getSavingsPlansUtilizationDetails_timePeriod - The time period that you want the usage and costs for. The Start date must be within 13 months. The End date must be after the Start date, and before the current date. Future dates can't be used as an End date.

Request Lenses

getSavingsPlansUtilizationDetails_nextToken :: Lens' GetSavingsPlansUtilizationDetails (Maybe Text)

The token to retrieve the next set of results. Amazon Web Services provides the token when the response from a previous call has more results than the maximum page size.

getSavingsPlansUtilizationDetails_dataType :: Lens' GetSavingsPlansUtilizationDetails (Maybe [SavingsPlansDataType])

The data type.

getSavingsPlansUtilizationDetails_filter :: Lens' GetSavingsPlansUtilizationDetails (Maybe Expression)

Filters Savings Plans utilization coverage data for active Savings Plans dimensions. You can filter data with the following dimensions:

LINKED_ACCOUNT

SAVINGS_PLAN_ARN

REGION

PAYMENT_OPTION

INSTANCE_TYPE_FAMILY

GetSavingsPlansUtilizationDetails uses the same Expression object as the other operations, but only AND is supported among each dimension.

getSavingsPlansUtilizationDetails_maxResults :: Lens' GetSavingsPlansUtilizationDetails (Maybe Natural)

The number of items to be returned in a response. The default is 20, with a minimum value of 1.

getSavingsPlansUtilizationDetails_sortBy :: Lens' GetSavingsPlansUtilizationDetails (Maybe SortDefinition)

The value by which you want to sort the data.

The following values are supported for Key:

UtilizationPercentage

TotalCommitment

UsedCommitment

UnusedCommitment

NetSavings

AmortizedRecurringCommitment

AmortizedUpfrontCommitment

Supported values for SortOrder are ASCENDING or DESCENDING.

getSavingsPlansUtilizationDetails_timePeriod :: Lens' GetSavingsPlansUtilizationDetails DateInterval

The time period that you want the usage and costs for. The Start date must be within 13 months. The End date must be after the Start date, and before the current date. Future dates can't be used as an End date.

Destructuring the Response

data GetSavingsPlansUtilizationDetailsResponse

See: newGetSavingsPlansUtilizationDetailsResponse smart constructor.

Constructors

GetSavingsPlansUtilizationDetailsResponse'

Fields

nextToken :: Maybe Text

The token to retrieve the next set of results. Amazon Web Services provides the token when the response from a previous call has more results than the maximum page size.

total :: Maybe SavingsPlansUtilizationAggregates

The total Savings Plans utilization, regardless of time period.

httpStatus :: Int

The response's http status code.

savingsPlansUtilizationDetails :: [SavingsPlansUtilizationDetail]

Retrieves a single daily or monthly Savings Plans utilization rate and details for your account.

timePeriod :: DateInterval

newGetSavingsPlansUtilizationDetailsResponse

Arguments

:: Int
$sel:httpStatus:GetSavingsPlansUtilizationDetailsResponse'

-> DateInterval
$sel:timePeriod:GetSavingsPlansUtilizationDetails'

-> GetSavingsPlansUtilizationDetailsResponse

Create a value of GetSavingsPlansUtilizationDetailsResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:nextToken:GetSavingsPlansUtilizationDetails', getSavingsPlansUtilizationDetailsResponse_nextToken - The token to retrieve the next set of results. Amazon Web Services provides the token when the response from a previous call has more results than the maximum page size.

$sel:total:GetSavingsPlansUtilizationDetailsResponse', getSavingsPlansUtilizationDetailsResponse_total - The total Savings Plans utilization, regardless of time period.

$sel:httpStatus:GetSavingsPlansUtilizationDetailsResponse', getSavingsPlansUtilizationDetailsResponse_httpStatus - The response's http status code.

$sel:savingsPlansUtilizationDetails:GetSavingsPlansUtilizationDetailsResponse', getSavingsPlansUtilizationDetailsResponse_savingsPlansUtilizationDetails - Retrieves a single daily or monthly Savings Plans utilization rate and details for your account.

$sel:timePeriod:GetSavingsPlansUtilizationDetails', getSavingsPlansUtilizationDetailsResponse_timePeriod - Undocumented member.

Response Lenses

getSavingsPlansUtilizationDetailsResponse_nextToken :: Lens' GetSavingsPlansUtilizationDetailsResponse (Maybe Text)

The token to retrieve the next set of results. Amazon Web Services provides the token when the response from a previous call has more results than the maximum page size.

getSavingsPlansUtilizationDetailsResponse_total :: Lens' GetSavingsPlansUtilizationDetailsResponse (Maybe SavingsPlansUtilizationAggregates)

The total Savings Plans utilization, regardless of time period.

getSavingsPlansUtilizationDetailsResponse_httpStatus :: Lens' GetSavingsPlansUtilizationDetailsResponse Int

The response's http status code.

getSavingsPlansUtilizationDetailsResponse_savingsPlansUtilizationDetails :: Lens' GetSavingsPlansUtilizationDetailsResponse [SavingsPlansUtilizationDetail]

Retrieves a single daily or monthly Savings Plans utilization rate and details for your account.

getSavingsPlansUtilizationDetailsResponse_timePeriod :: Lens' GetSavingsPlansUtilizationDetailsResponse DateInterval

Undocumented member.