Amazonka.CostExplorer.GetUsageForecast

Description

Retrieves a forecast for how much Amazon Web Services predicts that you will use over the forecast time period that you select, based on your past usage.

Creating a Request

data GetUsageForecast

See: newGetUsageForecast smart constructor.

Constructors

GetUsageForecast'

Fields

predictionIntervalLevel :: Maybe Natural

Cost Explorer always returns the mean forecast as a single point. You can request a prediction interval around the mean by specifying a confidence level. The higher the confidence level, the more confident Cost Explorer is about the actual value falling in the prediction interval. Higher confidence levels result in wider prediction intervals.

filter' :: Maybe Expression

The filters that you want to use to filter your forecast. The GetUsageForecast API supports filtering by the following dimensions:

AZ

INSTANCE_TYPE

LINKED_ACCOUNT

LINKED_ACCOUNT_NAME

OPERATION

PURCHASE_TYPE

REGION

SERVICE

USAGE_TYPE

USAGE_TYPE_GROUP

RECORD_TYPE

OPERATING_SYSTEM

TENANCY

SCOPE

PLATFORM

SUBSCRIPTION_ID

LEGAL_ENTITY_NAME

DEPLOYMENT_OPTION

DATABASE_ENGINE

INSTANCE_TYPE_FAMILY

BILLING_ENTITY

RESERVATION_ID

SAVINGS_PLAN_ARN

timePeriod :: DateInterval

The start and end dates of the period that you want to retrieve usage forecast for. The start date is inclusive, but the end date is exclusive. For example, if start is 2017-01-01 and end is 2017-05-01, then the cost and usage data is retrieved from 2017-01-01 up to and including 2017-04-30 but not including 2017-05-01. The start date must be equal to or later than the current date to avoid a validation error.

metric :: Metric

Which metric Cost Explorer uses to create your forecast.

Valid values for a GetUsageForecast call are the following:

USAGE_QUANTITY

NORMALIZED_USAGE_AMOUNT

granularity :: Granularity

How granular you want the forecast to be. You can get 3 months of DAILY forecasts or 12 months of MONTHLY forecasts.

The GetUsageForecast operation supports only DAILY and MONTHLY granularities.

newGetUsageForecast

Arguments

:: DateInterval
$sel:timePeriod:GetUsageForecast'

-> Metric
$sel:metric:GetUsageForecast'

-> Granularity
$sel:granularity:GetUsageForecast'

-> GetUsageForecast

Create a value of GetUsageForecast with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:predictionIntervalLevel:GetUsageForecast', getUsageForecast_predictionIntervalLevel - Cost Explorer always returns the mean forecast as a single point. You can request a prediction interval around the mean by specifying a confidence level. The higher the confidence level, the more confident Cost Explorer is about the actual value falling in the prediction interval. Higher confidence levels result in wider prediction intervals.

$sel:filter':GetUsageForecast', getUsageForecast_filter - The filters that you want to use to filter your forecast. The GetUsageForecast API supports filtering by the following dimensions:

AZ

INSTANCE_TYPE

LINKED_ACCOUNT

LINKED_ACCOUNT_NAME

OPERATION

PURCHASE_TYPE

REGION

SERVICE

USAGE_TYPE

USAGE_TYPE_GROUP

RECORD_TYPE

OPERATING_SYSTEM

TENANCY

SCOPE

PLATFORM

SUBSCRIPTION_ID

LEGAL_ENTITY_NAME

DEPLOYMENT_OPTION

DATABASE_ENGINE

INSTANCE_TYPE_FAMILY

BILLING_ENTITY

RESERVATION_ID

SAVINGS_PLAN_ARN

$sel:timePeriod:GetUsageForecast', getUsageForecast_timePeriod - The start and end dates of the period that you want to retrieve usage forecast for. The start date is inclusive, but the end date is exclusive. For example, if start is 2017-01-01 and end is 2017-05-01, then the cost and usage data is retrieved from 2017-01-01 up to and including 2017-04-30 but not including 2017-05-01. The start date must be equal to or later than the current date to avoid a validation error.

$sel:metric:GetUsageForecast', getUsageForecast_metric - Which metric Cost Explorer uses to create your forecast.

Valid values for a GetUsageForecast call are the following:

USAGE_QUANTITY

NORMALIZED_USAGE_AMOUNT

$sel:granularity:GetUsageForecast', getUsageForecast_granularity - How granular you want the forecast to be. You can get 3 months of DAILY forecasts or 12 months of MONTHLY forecasts.

The GetUsageForecast operation supports only DAILY and MONTHLY granularities.

Request Lenses

getUsageForecast_predictionIntervalLevel :: Lens' GetUsageForecast (Maybe Natural)

Cost Explorer always returns the mean forecast as a single point. You can request a prediction interval around the mean by specifying a confidence level. The higher the confidence level, the more confident Cost Explorer is about the actual value falling in the prediction interval. Higher confidence levels result in wider prediction intervals.

getUsageForecast_filter :: Lens' GetUsageForecast (Maybe Expression)

The filters that you want to use to filter your forecast. The GetUsageForecast API supports filtering by the following dimensions:

AZ

INSTANCE_TYPE

LINKED_ACCOUNT

LINKED_ACCOUNT_NAME

OPERATION

PURCHASE_TYPE

REGION

SERVICE

USAGE_TYPE

USAGE_TYPE_GROUP

RECORD_TYPE

OPERATING_SYSTEM

TENANCY

SCOPE

PLATFORM

SUBSCRIPTION_ID

LEGAL_ENTITY_NAME

DEPLOYMENT_OPTION

DATABASE_ENGINE

INSTANCE_TYPE_FAMILY

BILLING_ENTITY

RESERVATION_ID

SAVINGS_PLAN_ARN

getUsageForecast_timePeriod :: Lens' GetUsageForecast DateInterval

The start and end dates of the period that you want to retrieve usage forecast for. The start date is inclusive, but the end date is exclusive. For example, if start is 2017-01-01 and end is 2017-05-01, then the cost and usage data is retrieved from 2017-01-01 up to and including 2017-04-30 but not including 2017-05-01. The start date must be equal to or later than the current date to avoid a validation error.

getUsageForecast_metric :: Lens' GetUsageForecast Metric

Which metric Cost Explorer uses to create your forecast.

Valid values for a GetUsageForecast call are the following:

USAGE_QUANTITY

NORMALIZED_USAGE_AMOUNT

getUsageForecast_granularity :: Lens' GetUsageForecast Granularity

How granular you want the forecast to be. You can get 3 months of DAILY forecasts or 12 months of MONTHLY forecasts.

The GetUsageForecast operation supports only DAILY and MONTHLY granularities.

Destructuring the Response

data GetUsageForecastResponse

See: newGetUsageForecastResponse smart constructor.

Constructors

GetUsageForecastResponse'

Fields

forecastResultsByTime :: Maybe [ForecastResult]

The forecasts for your query, in order. For DAILY forecasts, this is a list of days. For MONTHLY forecasts, this is a list of months.

total :: Maybe MetricValue

How much you're forecasted to use over the forecast period.

httpStatus :: Int

The response's http status code.

newGetUsageForecastResponse

Arguments

:: Int
$sel:httpStatus:GetUsageForecastResponse'

-> GetUsageForecastResponse

Create a value of GetUsageForecastResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:forecastResultsByTime:GetUsageForecastResponse', getUsageForecastResponse_forecastResultsByTime - The forecasts for your query, in order. For DAILY forecasts, this is a list of days. For MONTHLY forecasts, this is a list of months.

$sel:total:GetUsageForecastResponse', getUsageForecastResponse_total - How much you're forecasted to use over the forecast period.

$sel:httpStatus:GetUsageForecastResponse', getUsageForecastResponse_httpStatus - The response's http status code.

Response Lenses

getUsageForecastResponse_forecastResultsByTime :: Lens' GetUsageForecastResponse (Maybe [ForecastResult])

The forecasts for your query, in order. For DAILY forecasts, this is a list of days. For MONTHLY forecasts, this is a list of months.

getUsageForecastResponse_total :: Lens' GetUsageForecastResponse (Maybe MetricValue)

How much you're forecasted to use over the forecast period.

getUsageForecastResponse_httpStatus :: Lens' GetUsageForecastResponse Int

The response's http status code.