Amazonka.CostExplorer.Types.AnomalyMonitor

Description

Documentation

data AnomalyMonitor

This object continuously inspects your account's cost data for anomalies. It's based on MonitorType and MonitorSpecification. The content consists of detailed metadata and the current status of the monitor object.

See: newAnomalyMonitor smart constructor.

Constructors

AnomalyMonitor'

Fields

dimensionalValueCount :: Maybe Natural

The value for evaluated dimensions.

monitorSpecification :: Maybe Expression

monitorDimension :: Maybe MonitorDimension

The dimensions to evaluate.

creationDate :: Maybe Text

The date when the monitor was created.

lastUpdatedDate :: Maybe Text

The date when the monitor was last updated.

lastEvaluatedDate :: Maybe Text

The date when the monitor last evaluated for anomalies.

monitorArn :: Maybe Text

The Amazon Resource Name (ARN) value.

monitorName :: Text

The name of the monitor.

monitorType :: MonitorType

The possible type values.

newAnomalyMonitor

Arguments

:: Text
$sel:monitorName:AnomalyMonitor'

-> MonitorType
$sel:monitorType:AnomalyMonitor'

-> AnomalyMonitor

Create a value of AnomalyMonitor with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:dimensionalValueCount:AnomalyMonitor', anomalyMonitor_dimensionalValueCount - The value for evaluated dimensions.

$sel:monitorSpecification:AnomalyMonitor', anomalyMonitor_monitorSpecification - Undocumented member.

$sel:monitorDimension:AnomalyMonitor', anomalyMonitor_monitorDimension - The dimensions to evaluate.

$sel:creationDate:AnomalyMonitor', anomalyMonitor_creationDate - The date when the monitor was created.

$sel:lastUpdatedDate:AnomalyMonitor', anomalyMonitor_lastUpdatedDate - The date when the monitor was last updated.

$sel:lastEvaluatedDate:AnomalyMonitor', anomalyMonitor_lastEvaluatedDate - The date when the monitor last evaluated for anomalies.

$sel:monitorArn:AnomalyMonitor', anomalyMonitor_monitorArn - The Amazon Resource Name (ARN) value.

$sel:monitorName:AnomalyMonitor', anomalyMonitor_monitorName - The name of the monitor.

$sel:monitorType:AnomalyMonitor', anomalyMonitor_monitorType - The possible type values.

anomalyMonitor_dimensionalValueCount :: Lens' AnomalyMonitor (Maybe Natural)

The value for evaluated dimensions.

anomalyMonitor_monitorSpecification :: Lens' AnomalyMonitor (Maybe Expression)

Undocumented member.

anomalyMonitor_monitorDimension :: Lens' AnomalyMonitor (Maybe MonitorDimension)

The dimensions to evaluate.

anomalyMonitor_creationDate :: Lens' AnomalyMonitor (Maybe Text)

The date when the monitor was created.

anomalyMonitor_lastUpdatedDate :: Lens' AnomalyMonitor (Maybe Text)

The date when the monitor was last updated.

anomalyMonitor_lastEvaluatedDate :: Lens' AnomalyMonitor (Maybe Text)

The date when the monitor last evaluated for anomalies.

anomalyMonitor_monitorArn :: Lens' AnomalyMonitor (Maybe Text)

The Amazon Resource Name (ARN) value.

anomalyMonitor_monitorName :: Lens' AnomalyMonitor Text

The name of the monitor.

anomalyMonitor_monitorType :: Lens' AnomalyMonitor MonitorType

The possible type values.