Amazonka.CostExplorer.Types.CostCategorySplitChargeRule

Description

Documentation

data CostCategorySplitChargeRule

Use the split charge rule to split the cost of one Cost Category value across several other target values.

See: newCostCategorySplitChargeRule smart constructor.

Constructors

CostCategorySplitChargeRule'

Fields

parameters :: Maybe (NonEmpty CostCategorySplitChargeRuleParameter)

The parameters for a split charge method. This is only required for the FIXED method.

source :: Text

The Cost Category value that you want to split. That value can't be used as a source or a target in other split charge rules. To indicate uncategorized costs, you can use an empty string as the source.

targets :: NonEmpty Text

The Cost Category values that you want to split costs across. These values can't be used as a source in other split charge rules.

method :: CostCategorySplitChargeMethod

The method that's used to define how to split your source costs across your targets.

Proportional - Allocates charges across your targets based on the proportional weighted cost of each target.

Fixed - Allocates charges across your targets based on your defined allocation percentage.

@Even@ - Allocates costs evenly across all targets.

newCostCategorySplitChargeRule

Arguments

:: Text
$sel:source:CostCategorySplitChargeRule'

-> NonEmpty Text
$sel:targets:CostCategorySplitChargeRule'

-> CostCategorySplitChargeMethod
$sel:method:CostCategorySplitChargeRule'

-> CostCategorySplitChargeRule

Create a value of CostCategorySplitChargeRule with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:parameters:CostCategorySplitChargeRule', costCategorySplitChargeRule_parameters - The parameters for a split charge method. This is only required for the FIXED method.

$sel:source:CostCategorySplitChargeRule', costCategorySplitChargeRule_source - The Cost Category value that you want to split. That value can't be used as a source or a target in other split charge rules. To indicate uncategorized costs, you can use an empty string as the source.

$sel:targets:CostCategorySplitChargeRule', costCategorySplitChargeRule_targets - The Cost Category values that you want to split costs across. These values can't be used as a source in other split charge rules.

$sel:method:CostCategorySplitChargeRule', costCategorySplitChargeRule_method - The method that's used to define how to split your source costs across your targets.

Proportional - Allocates charges across your targets based on the proportional weighted cost of each target.

Fixed - Allocates charges across your targets based on your defined allocation percentage.

@Even@ - Allocates costs evenly across all targets.

costCategorySplitChargeRule_parameters :: Lens' CostCategorySplitChargeRule (Maybe (NonEmpty CostCategorySplitChargeRuleParameter))

The parameters for a split charge method. This is only required for the FIXED method.

costCategorySplitChargeRule_source :: Lens' CostCategorySplitChargeRule Text

The Cost Category value that you want to split. That value can't be used as a source or a target in other split charge rules. To indicate uncategorized costs, you can use an empty string as the source.

costCategorySplitChargeRule_targets :: Lens' CostCategorySplitChargeRule (NonEmpty Text)

The Cost Category values that you want to split costs across. These values can't be used as a source in other split charge rules.

costCategorySplitChargeRule_method :: Lens' CostCategorySplitChargeRule CostCategorySplitChargeMethod

The method that's used to define how to split your source costs across your targets.

Proportional - Allocates charges across your targets based on the proportional weighted cost of each target.

Fixed - Allocates charges across your targets based on your defined allocation percentage.

@Even@ - Allocates costs evenly across all targets.