Amazonka.CostExplorer.Types.ElastiCacheInstanceDetails

Description

Documentation

data ElastiCacheInstanceDetails

Details about the Amazon ElastiCache instances that Amazon Web Services recommends that you purchase.

See: newElastiCacheInstanceDetails smart constructor.

Constructors

ElastiCacheInstanceDetails'

Fields

currentGeneration :: Maybe Bool

Determines whether the recommendation is for a current generation instance.

productDescription :: Maybe Text

The description of the recommended reservation.

family :: Maybe Text

The instance family of the recommended reservation.

sizeFlexEligible :: Maybe Bool

Determines whether the recommended reservation is size flexible.

region :: Maybe Text

The Amazon Web Services Region of the recommended reservation.

nodeType :: Maybe Text

The type of node that Amazon Web Services recommends.

newElastiCacheInstanceDetails :: ElastiCacheInstanceDetails

Create a value of ElastiCacheInstanceDetails with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:currentGeneration:ElastiCacheInstanceDetails', elastiCacheInstanceDetails_currentGeneration - Determines whether the recommendation is for a current generation instance.

$sel:productDescription:ElastiCacheInstanceDetails', elastiCacheInstanceDetails_productDescription - The description of the recommended reservation.

$sel:family:ElastiCacheInstanceDetails', elastiCacheInstanceDetails_family - The instance family of the recommended reservation.

$sel:sizeFlexEligible:ElastiCacheInstanceDetails', elastiCacheInstanceDetails_sizeFlexEligible - Determines whether the recommended reservation is size flexible.

$sel:region:ElastiCacheInstanceDetails', elastiCacheInstanceDetails_region - The Amazon Web Services Region of the recommended reservation.

$sel:nodeType:ElastiCacheInstanceDetails', elastiCacheInstanceDetails_nodeType - The type of node that Amazon Web Services recommends.

elastiCacheInstanceDetails_currentGeneration :: Lens' ElastiCacheInstanceDetails (Maybe Bool)

Determines whether the recommendation is for a current generation instance.

elastiCacheInstanceDetails_productDescription :: Lens' ElastiCacheInstanceDetails (Maybe Text)

The description of the recommended reservation.

elastiCacheInstanceDetails_family :: Lens' ElastiCacheInstanceDetails (Maybe Text)

The instance family of the recommended reservation.

elastiCacheInstanceDetails_sizeFlexEligible :: Lens' ElastiCacheInstanceDetails (Maybe Bool)

Determines whether the recommended reservation is size flexible.

elastiCacheInstanceDetails_region :: Lens' ElastiCacheInstanceDetails (Maybe Text)

The Amazon Web Services Region of the recommended reservation.

elastiCacheInstanceDetails_nodeType :: Lens' ElastiCacheInstanceDetails (Maybe Text)

The type of node that Amazon Web Services recommends.