Amazonka.CostExplorer.Types.RDSInstanceDetails

Description

Documentation

data RDSInstanceDetails

Details about the Amazon RDS instances that Amazon Web Services recommends that you purchase.

See: newRDSInstanceDetails smart constructor.

Constructors

RDSInstanceDetails'

Fields

currentGeneration :: Maybe Bool

Determines whether the recommendation is for a current-generation instance.

deploymentOption :: Maybe Text

Determines whether the recommendation is for a reservation in a single Availability Zone or a reservation with a backup in a second Availability Zone.

family :: Maybe Text

The instance family of the recommended reservation.

instanceType :: Maybe Text

The type of instance that Amazon Web Services recommends.

licenseModel :: Maybe Text

The license model that the recommended reservation supports.

sizeFlexEligible :: Maybe Bool

Determines whether the recommended reservation is size flexible.

region :: Maybe Text

The Amazon Web Services Region of the recommended reservation.

databaseEngine :: Maybe Text

The database engine that the recommended reservation supports.

databaseEdition :: Maybe Text

The database edition that the recommended reservation supports.

newRDSInstanceDetails :: RDSInstanceDetails

Create a value of RDSInstanceDetails with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:currentGeneration:RDSInstanceDetails', rDSInstanceDetails_currentGeneration - Determines whether the recommendation is for a current-generation instance.

$sel:deploymentOption:RDSInstanceDetails', rDSInstanceDetails_deploymentOption - Determines whether the recommendation is for a reservation in a single Availability Zone or a reservation with a backup in a second Availability Zone.

$sel:family:RDSInstanceDetails', rDSInstanceDetails_family - The instance family of the recommended reservation.

$sel:instanceType:RDSInstanceDetails', rDSInstanceDetails_instanceType - The type of instance that Amazon Web Services recommends.

$sel:licenseModel:RDSInstanceDetails', rDSInstanceDetails_licenseModel - The license model that the recommended reservation supports.

$sel:sizeFlexEligible:RDSInstanceDetails', rDSInstanceDetails_sizeFlexEligible - Determines whether the recommended reservation is size flexible.

$sel:region:RDSInstanceDetails', rDSInstanceDetails_region - The Amazon Web Services Region of the recommended reservation.

$sel:databaseEngine:RDSInstanceDetails', rDSInstanceDetails_databaseEngine - The database engine that the recommended reservation supports.

$sel:databaseEdition:RDSInstanceDetails', rDSInstanceDetails_databaseEdition - The database edition that the recommended reservation supports.

rDSInstanceDetails_currentGeneration :: Lens' RDSInstanceDetails (Maybe Bool)

Determines whether the recommendation is for a current-generation instance.

rDSInstanceDetails_deploymentOption :: Lens' RDSInstanceDetails (Maybe Text)

Determines whether the recommendation is for a reservation in a single Availability Zone or a reservation with a backup in a second Availability Zone.

rDSInstanceDetails_family :: Lens' RDSInstanceDetails (Maybe Text)

The instance family of the recommended reservation.

rDSInstanceDetails_instanceType :: Lens' RDSInstanceDetails (Maybe Text)

The type of instance that Amazon Web Services recommends.

rDSInstanceDetails_licenseModel :: Lens' RDSInstanceDetails (Maybe Text)

The license model that the recommended reservation supports.

rDSInstanceDetails_sizeFlexEligible :: Lens' RDSInstanceDetails (Maybe Bool)

Determines whether the recommended reservation is size flexible.

rDSInstanceDetails_region :: Lens' RDSInstanceDetails (Maybe Text)

The Amazon Web Services Region of the recommended reservation.

rDSInstanceDetails_databaseEngine :: Lens' RDSInstanceDetails (Maybe Text)

The database engine that the recommended reservation supports.

rDSInstanceDetails_databaseEdition :: Lens' RDSInstanceDetails (Maybe Text)

The database edition that the recommended reservation supports.