Amazonka.CostExplorer.Types.ReservationAggregates

Description

Documentation

data ReservationAggregates

The aggregated numbers for your reservation usage.

See: newReservationAggregates smart constructor.

Constructors

ReservationAggregates'

Fields

purchasedHours :: Maybe Text

How many reservation hours that you purchased.

totalActualHours :: Maybe Text

The total number of reservation hours that you used.

utilizationPercentage :: Maybe Text

The percentage of reservation time that you used.

totalAmortizedFee :: Maybe Text

The total cost of your reservation. It's amortized over the reservation period.

unusedUnits :: Maybe Text

The number of Amazon EC2 reservation hours that you didn't use. It's converted to normalized units. Normalized units are available only for Amazon EC2 usage after November 11, 2017.

unrealizedSavings :: Maybe Text

The unrealized savings because of purchasing and using a reservation.

rICostForUnusedHours :: Maybe Text

The cost of unused hours for your reservation.

unusedHours :: Maybe Text

The number of reservation hours that you didn't use.

realizedSavings :: Maybe Text

The realized savings because of purchasing and using a reservation.

purchasedUnits :: Maybe Text

The number of Amazon EC2 reservation hours that you purchased. It's converted to normalized units. Normalized units are available only for Amazon EC2 usage after November 11, 2017.

amortizedUpfrontFee :: Maybe Text

The upfront cost of your reservation. It's amortized over the reservation period.

amortizedRecurringFee :: Maybe Text

The monthly cost of your reservation. It's amortized over the reservation period.

utilizationPercentageInUnits :: Maybe Text

The percentage of Amazon EC2 reservation time that you used. It's converted to normalized units. Normalized units are available only for Amazon EC2 usage after November 11, 2017.

netRISavings :: Maybe Text

How much you saved due to purchasing and utilizing reservation. Amazon Web Services calculates this by subtracting TotalAmortizedFee from OnDemandCostOfRIHoursUsed.

onDemandCostOfRIHoursUsed :: Maybe Text

How much your reservation costs if charged On-Demand rates.

totalPotentialRISavings :: Maybe Text

How much you might save if you use your entire reservation.

totalActualUnits :: Maybe Text

The total number of Amazon EC2 reservation hours that you used. It's converted to normalized units. Normalized units are available only for Amazon EC2 usage after November 11, 2017.

newReservationAggregates :: ReservationAggregates

Create a value of ReservationAggregates with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:purchasedHours:ReservationAggregates', reservationAggregates_purchasedHours - How many reservation hours that you purchased.

$sel:totalActualHours:ReservationAggregates', reservationAggregates_totalActualHours - The total number of reservation hours that you used.

$sel:utilizationPercentage:ReservationAggregates', reservationAggregates_utilizationPercentage - The percentage of reservation time that you used.

$sel:totalAmortizedFee:ReservationAggregates', reservationAggregates_totalAmortizedFee - The total cost of your reservation. It's amortized over the reservation period.

$sel:unusedUnits:ReservationAggregates', reservationAggregates_unusedUnits - The number of Amazon EC2 reservation hours that you didn't use. It's converted to normalized units. Normalized units are available only for Amazon EC2 usage after November 11, 2017.

$sel:unrealizedSavings:ReservationAggregates', reservationAggregates_unrealizedSavings - The unrealized savings because of purchasing and using a reservation.

$sel:rICostForUnusedHours:ReservationAggregates', reservationAggregates_rICostForUnusedHours - The cost of unused hours for your reservation.

$sel:unusedHours:ReservationAggregates', reservationAggregates_unusedHours - The number of reservation hours that you didn't use.

$sel:realizedSavings:ReservationAggregates', reservationAggregates_realizedSavings - The realized savings because of purchasing and using a reservation.

$sel:purchasedUnits:ReservationAggregates', reservationAggregates_purchasedUnits - The number of Amazon EC2 reservation hours that you purchased. It's converted to normalized units. Normalized units are available only for Amazon EC2 usage after November 11, 2017.

$sel:amortizedUpfrontFee:ReservationAggregates', reservationAggregates_amortizedUpfrontFee - The upfront cost of your reservation. It's amortized over the reservation period.

$sel:amortizedRecurringFee:ReservationAggregates', reservationAggregates_amortizedRecurringFee - The monthly cost of your reservation. It's amortized over the reservation period.

$sel:utilizationPercentageInUnits:ReservationAggregates', reservationAggregates_utilizationPercentageInUnits - The percentage of Amazon EC2 reservation time that you used. It's converted to normalized units. Normalized units are available only for Amazon EC2 usage after November 11, 2017.

$sel:netRISavings:ReservationAggregates', reservationAggregates_netRISavings - How much you saved due to purchasing and utilizing reservation. Amazon Web Services calculates this by subtracting TotalAmortizedFee from OnDemandCostOfRIHoursUsed.

$sel:onDemandCostOfRIHoursUsed:ReservationAggregates', reservationAggregates_onDemandCostOfRIHoursUsed - How much your reservation costs if charged On-Demand rates.

$sel:totalPotentialRISavings:ReservationAggregates', reservationAggregates_totalPotentialRISavings - How much you might save if you use your entire reservation.

$sel:totalActualUnits:ReservationAggregates', reservationAggregates_totalActualUnits - The total number of Amazon EC2 reservation hours that you used. It's converted to normalized units. Normalized units are available only for Amazon EC2 usage after November 11, 2017.

reservationAggregates_purchasedHours :: Lens' ReservationAggregates (Maybe Text)

How many reservation hours that you purchased.

reservationAggregates_totalActualHours :: Lens' ReservationAggregates (Maybe Text)

The total number of reservation hours that you used.

reservationAggregates_utilizationPercentage :: Lens' ReservationAggregates (Maybe Text)

The percentage of reservation time that you used.

reservationAggregates_totalAmortizedFee :: Lens' ReservationAggregates (Maybe Text)

The total cost of your reservation. It's amortized over the reservation period.

reservationAggregates_unusedUnits :: Lens' ReservationAggregates (Maybe Text)

The number of Amazon EC2 reservation hours that you didn't use. It's converted to normalized units. Normalized units are available only for Amazon EC2 usage after November 11, 2017.

reservationAggregates_unrealizedSavings :: Lens' ReservationAggregates (Maybe Text)

The unrealized savings because of purchasing and using a reservation.

reservationAggregates_rICostForUnusedHours :: Lens' ReservationAggregates (Maybe Text)

The cost of unused hours for your reservation.

reservationAggregates_unusedHours :: Lens' ReservationAggregates (Maybe Text)

The number of reservation hours that you didn't use.

reservationAggregates_realizedSavings :: Lens' ReservationAggregates (Maybe Text)

The realized savings because of purchasing and using a reservation.

reservationAggregates_purchasedUnits :: Lens' ReservationAggregates (Maybe Text)

The number of Amazon EC2 reservation hours that you purchased. It's converted to normalized units. Normalized units are available only for Amazon EC2 usage after November 11, 2017.

reservationAggregates_amortizedUpfrontFee :: Lens' ReservationAggregates (Maybe Text)

The upfront cost of your reservation. It's amortized over the reservation period.

reservationAggregates_amortizedRecurringFee :: Lens' ReservationAggregates (Maybe Text)

The monthly cost of your reservation. It's amortized over the reservation period.

reservationAggregates_utilizationPercentageInUnits :: Lens' ReservationAggregates (Maybe Text)

The percentage of Amazon EC2 reservation time that you used. It's converted to normalized units. Normalized units are available only for Amazon EC2 usage after November 11, 2017.

reservationAggregates_netRISavings :: Lens' ReservationAggregates (Maybe Text)

How much you saved due to purchasing and utilizing reservation. Amazon Web Services calculates this by subtracting TotalAmortizedFee from OnDemandCostOfRIHoursUsed.

reservationAggregates_onDemandCostOfRIHoursUsed :: Lens' ReservationAggregates (Maybe Text)

How much your reservation costs if charged On-Demand rates.

reservationAggregates_totalPotentialRISavings :: Lens' ReservationAggregates (Maybe Text)

How much you might save if you use your entire reservation.

reservationAggregates_totalActualUnits :: Lens' ReservationAggregates (Maybe Text)

The total number of Amazon EC2 reservation hours that you used. It's converted to normalized units. Normalized units are available only for Amazon EC2 usage after November 11, 2017.