Amazonka.CostExplorer.Types.ReservationPurchaseRecommendationDetail

Description

Documentation

data ReservationPurchaseRecommendationDetail

Details about your recommended reservation purchase.

See: newReservationPurchaseRecommendationDetail smart constructor.

Constructors

ReservationPurchaseRecommendationDetail'

Fields

maximumNormalizedUnitsUsedPerHour :: Maybe Text

The maximum number of normalized units that you used in an hour during the historical period. Amazon Web Services uses this to calculate your recommended reservation purchases.

recurringStandardMonthlyCost :: Maybe Text

How much purchasing this instance costs you on a monthly basis.

averageNormalizedUnitsUsedPerHour :: Maybe Text

The average number of normalized units that you used in an hour during the historical period. Amazon Web Services uses this to calculate your recommended reservation purchases.

currencyCode :: Maybe Text

The currency code that Amazon Web Services used to calculate the costs for this instance.

estimatedMonthlySavingsPercentage :: Maybe Text

How much Amazon Web Services estimates that this specific recommendation could save you in a month, as a percentage of your overall costs.

recommendedNormalizedUnitsToPurchase :: Maybe Text

The number of normalized units that Amazon Web Services recommends that you purchase.

averageUtilization :: Maybe Text

The average utilization of your instances. Amazon Web Services uses this to calculate your recommended reservation purchases.

accountId :: Maybe Text

The account that this RI recommendation is for.

estimatedMonthlySavingsAmount :: Maybe Text

How much Amazon Web Services estimates that this specific recommendation could save you in a month.

upfrontCost :: Maybe Text

How much purchasing this instance costs you upfront.

minimumNormalizedUnitsUsedPerHour :: Maybe Text

The minimum number of normalized units that you used in an hour during the historical period. Amazon Web Services uses this to calculate your recommended reservation purchases.

estimatedMonthlyOnDemandCost :: Maybe Text

How much Amazon Web Services estimates that you spend on On-Demand Instances in a month.

recommendedNumberOfInstancesToPurchase :: Maybe Text

The number of instances that Amazon Web Services recommends that you purchase.

maximumNumberOfInstancesUsedPerHour :: Maybe Text

The maximum number of instances that you used in an hour during the historical period. Amazon Web Services uses this to calculate your recommended reservation purchases.

estimatedReservationCostForLookbackPeriod :: Maybe Text

How much Amazon Web Services estimates that you would have spent for all usage during the specified historical period if you had a reservation.

instanceDetails :: Maybe InstanceDetails

Details about the instances that Amazon Web Services recommends that you purchase.

averageNumberOfInstancesUsedPerHour :: Maybe Text

The average number of instances that you used in an hour during the historical period. Amazon Web Services uses this to calculate your recommended reservation purchases.

minimumNumberOfInstancesUsedPerHour :: Maybe Text

The minimum number of instances that you used in an hour during the historical period. Amazon Web Services uses this to calculate your recommended reservation purchases.

estimatedBreakEvenInMonths :: Maybe Text

How long Amazon Web Services estimates that it takes for this instance to start saving you money, in months.

newReservationPurchaseRecommendationDetail :: ReservationPurchaseRecommendationDetail

Create a value of ReservationPurchaseRecommendationDetail with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:maximumNormalizedUnitsUsedPerHour:ReservationPurchaseRecommendationDetail', reservationPurchaseRecommendationDetail_maximumNormalizedUnitsUsedPerHour - The maximum number of normalized units that you used in an hour during the historical period. Amazon Web Services uses this to calculate your recommended reservation purchases.

$sel:recurringStandardMonthlyCost:ReservationPurchaseRecommendationDetail', reservationPurchaseRecommendationDetail_recurringStandardMonthlyCost - How much purchasing this instance costs you on a monthly basis.

$sel:averageNormalizedUnitsUsedPerHour:ReservationPurchaseRecommendationDetail', reservationPurchaseRecommendationDetail_averageNormalizedUnitsUsedPerHour - The average number of normalized units that you used in an hour during the historical period. Amazon Web Services uses this to calculate your recommended reservation purchases.

$sel:currencyCode:ReservationPurchaseRecommendationDetail', reservationPurchaseRecommendationDetail_currencyCode - The currency code that Amazon Web Services used to calculate the costs for this instance.

$sel:estimatedMonthlySavingsPercentage:ReservationPurchaseRecommendationDetail', reservationPurchaseRecommendationDetail_estimatedMonthlySavingsPercentage - How much Amazon Web Services estimates that this specific recommendation could save you in a month, as a percentage of your overall costs.

$sel:recommendedNormalizedUnitsToPurchase:ReservationPurchaseRecommendationDetail', reservationPurchaseRecommendationDetail_recommendedNormalizedUnitsToPurchase - The number of normalized units that Amazon Web Services recommends that you purchase.

$sel:averageUtilization:ReservationPurchaseRecommendationDetail', reservationPurchaseRecommendationDetail_averageUtilization - The average utilization of your instances. Amazon Web Services uses this to calculate your recommended reservation purchases.

$sel:accountId:ReservationPurchaseRecommendationDetail', reservationPurchaseRecommendationDetail_accountId - The account that this RI recommendation is for.

$sel:estimatedMonthlySavingsAmount:ReservationPurchaseRecommendationDetail', reservationPurchaseRecommendationDetail_estimatedMonthlySavingsAmount - How much Amazon Web Services estimates that this specific recommendation could save you in a month.

$sel:upfrontCost:ReservationPurchaseRecommendationDetail', reservationPurchaseRecommendationDetail_upfrontCost - How much purchasing this instance costs you upfront.

$sel:minimumNormalizedUnitsUsedPerHour:ReservationPurchaseRecommendationDetail', reservationPurchaseRecommendationDetail_minimumNormalizedUnitsUsedPerHour - The minimum number of normalized units that you used in an hour during the historical period. Amazon Web Services uses this to calculate your recommended reservation purchases.

$sel:estimatedMonthlyOnDemandCost:ReservationPurchaseRecommendationDetail', reservationPurchaseRecommendationDetail_estimatedMonthlyOnDemandCost - How much Amazon Web Services estimates that you spend on On-Demand Instances in a month.

$sel:recommendedNumberOfInstancesToPurchase:ReservationPurchaseRecommendationDetail', reservationPurchaseRecommendationDetail_recommendedNumberOfInstancesToPurchase - The number of instances that Amazon Web Services recommends that you purchase.

$sel:maximumNumberOfInstancesUsedPerHour:ReservationPurchaseRecommendationDetail', reservationPurchaseRecommendationDetail_maximumNumberOfInstancesUsedPerHour - The maximum number of instances that you used in an hour during the historical period. Amazon Web Services uses this to calculate your recommended reservation purchases.

$sel:estimatedReservationCostForLookbackPeriod:ReservationPurchaseRecommendationDetail', reservationPurchaseRecommendationDetail_estimatedReservationCostForLookbackPeriod - How much Amazon Web Services estimates that you would have spent for all usage during the specified historical period if you had a reservation.

$sel:instanceDetails:ReservationPurchaseRecommendationDetail', reservationPurchaseRecommendationDetail_instanceDetails - Details about the instances that Amazon Web Services recommends that you purchase.

$sel:averageNumberOfInstancesUsedPerHour:ReservationPurchaseRecommendationDetail', reservationPurchaseRecommendationDetail_averageNumberOfInstancesUsedPerHour - The average number of instances that you used in an hour during the historical period. Amazon Web Services uses this to calculate your recommended reservation purchases.

$sel:minimumNumberOfInstancesUsedPerHour:ReservationPurchaseRecommendationDetail', reservationPurchaseRecommendationDetail_minimumNumberOfInstancesUsedPerHour - The minimum number of instances that you used in an hour during the historical period. Amazon Web Services uses this to calculate your recommended reservation purchases.

$sel:estimatedBreakEvenInMonths:ReservationPurchaseRecommendationDetail', reservationPurchaseRecommendationDetail_estimatedBreakEvenInMonths - How long Amazon Web Services estimates that it takes for this instance to start saving you money, in months.

reservationPurchaseRecommendationDetail_maximumNormalizedUnitsUsedPerHour :: Lens' ReservationPurchaseRecommendationDetail (Maybe Text)

The maximum number of normalized units that you used in an hour during the historical period. Amazon Web Services uses this to calculate your recommended reservation purchases.

reservationPurchaseRecommendationDetail_recurringStandardMonthlyCost :: Lens' ReservationPurchaseRecommendationDetail (Maybe Text)

How much purchasing this instance costs you on a monthly basis.

reservationPurchaseRecommendationDetail_averageNormalizedUnitsUsedPerHour :: Lens' ReservationPurchaseRecommendationDetail (Maybe Text)

The average number of normalized units that you used in an hour during the historical period. Amazon Web Services uses this to calculate your recommended reservation purchases.

reservationPurchaseRecommendationDetail_currencyCode :: Lens' ReservationPurchaseRecommendationDetail (Maybe Text)

The currency code that Amazon Web Services used to calculate the costs for this instance.

reservationPurchaseRecommendationDetail_estimatedMonthlySavingsPercentage :: Lens' ReservationPurchaseRecommendationDetail (Maybe Text)

How much Amazon Web Services estimates that this specific recommendation could save you in a month, as a percentage of your overall costs.

reservationPurchaseRecommendationDetail_recommendedNormalizedUnitsToPurchase :: Lens' ReservationPurchaseRecommendationDetail (Maybe Text)

The number of normalized units that Amazon Web Services recommends that you purchase.

reservationPurchaseRecommendationDetail_averageUtilization :: Lens' ReservationPurchaseRecommendationDetail (Maybe Text)

The average utilization of your instances. Amazon Web Services uses this to calculate your recommended reservation purchases.

reservationPurchaseRecommendationDetail_accountId :: Lens' ReservationPurchaseRecommendationDetail (Maybe Text)

The account that this RI recommendation is for.

reservationPurchaseRecommendationDetail_estimatedMonthlySavingsAmount :: Lens' ReservationPurchaseRecommendationDetail (Maybe Text)

How much Amazon Web Services estimates that this specific recommendation could save you in a month.

reservationPurchaseRecommendationDetail_upfrontCost :: Lens' ReservationPurchaseRecommendationDetail (Maybe Text)

How much purchasing this instance costs you upfront.

reservationPurchaseRecommendationDetail_minimumNormalizedUnitsUsedPerHour :: Lens' ReservationPurchaseRecommendationDetail (Maybe Text)

The minimum number of normalized units that you used in an hour during the historical period. Amazon Web Services uses this to calculate your recommended reservation purchases.

reservationPurchaseRecommendationDetail_estimatedMonthlyOnDemandCost :: Lens' ReservationPurchaseRecommendationDetail (Maybe Text)

How much Amazon Web Services estimates that you spend on On-Demand Instances in a month.

reservationPurchaseRecommendationDetail_recommendedNumberOfInstancesToPurchase :: Lens' ReservationPurchaseRecommendationDetail (Maybe Text)

The number of instances that Amazon Web Services recommends that you purchase.

reservationPurchaseRecommendationDetail_maximumNumberOfInstancesUsedPerHour :: Lens' ReservationPurchaseRecommendationDetail (Maybe Text)

The maximum number of instances that you used in an hour during the historical period. Amazon Web Services uses this to calculate your recommended reservation purchases.

reservationPurchaseRecommendationDetail_estimatedReservationCostForLookbackPeriod :: Lens' ReservationPurchaseRecommendationDetail (Maybe Text)

How much Amazon Web Services estimates that you would have spent for all usage during the specified historical period if you had a reservation.

reservationPurchaseRecommendationDetail_instanceDetails :: Lens' ReservationPurchaseRecommendationDetail (Maybe InstanceDetails)

Details about the instances that Amazon Web Services recommends that you purchase.

reservationPurchaseRecommendationDetail_averageNumberOfInstancesUsedPerHour :: Lens' ReservationPurchaseRecommendationDetail (Maybe Text)

The average number of instances that you used in an hour during the historical period. Amazon Web Services uses this to calculate your recommended reservation purchases.

reservationPurchaseRecommendationDetail_minimumNumberOfInstancesUsedPerHour :: Lens' ReservationPurchaseRecommendationDetail (Maybe Text)

The minimum number of instances that you used in an hour during the historical period. Amazon Web Services uses this to calculate your recommended reservation purchases.

reservationPurchaseRecommendationDetail_estimatedBreakEvenInMonths :: Lens' ReservationPurchaseRecommendationDetail (Maybe Text)

How long Amazon Web Services estimates that it takes for this instance to start saving you money, in months.