Amazonka.CostExplorer.Types.RootCause

Description

Documentation

data RootCause

The combination of Amazon Web Services service, linked account, Region, and usage type where a cost anomaly is observed.

See: newRootCause smart constructor.

Constructors

RootCause'

Fields

service :: Maybe Text

The Amazon Web Services service name that's associated with the cost anomaly.

usageType :: Maybe Text

The UsageType value that's associated with the cost anomaly.

linkedAccount :: Maybe Text

The member account value that's associated with the cost anomaly.

region :: Maybe Text

The Amazon Web Services Region that's associated with the cost anomaly.

newRootCause :: RootCause

Create a value of RootCause with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:service:RootCause', rootCause_service - The Amazon Web Services service name that's associated with the cost anomaly.

$sel:usageType:RootCause', rootCause_usageType - The UsageType value that's associated with the cost anomaly.

$sel:linkedAccount:RootCause', rootCause_linkedAccount - The member account value that's associated with the cost anomaly.

$sel:region:RootCause', rootCause_region - The Amazon Web Services Region that's associated with the cost anomaly.

rootCause_service :: Lens' RootCause (Maybe Text)

The Amazon Web Services service name that's associated with the cost anomaly.

rootCause_usageType :: Lens' RootCause (Maybe Text)

The UsageType value that's associated with the cost anomaly.

rootCause_linkedAccount :: Lens' RootCause (Maybe Text)

The member account value that's associated with the cost anomaly.

rootCause_region :: Lens' RootCause (Maybe Text)

The Amazon Web Services Region that's associated with the cost anomaly.