Amazonka.CostExplorer.Types.SavingsPlansCoverageData

Description

Documentation

data SavingsPlansCoverageData

Specific coverage percentage, On-Demand costs, and spend covered by Savings Plans, and total Savings Plans costs for an account.

See: newSavingsPlansCoverageData smart constructor.

Constructors

SavingsPlansCoverageData'

Fields

onDemandCost :: Maybe Text

The cost of your Amazon Web Services usage at the public On-Demand rate.

spendCoveredBySavingsPlans :: Maybe Text

The amount of your Amazon Web Services usage that is covered by a Savings Plans.

coveragePercentage :: Maybe Text

The percentage of your existing Savings Plans covered usage, divided by all of your eligible Savings Plans usage in an account (or set of accounts).

totalCost :: Maybe Text

The total cost of your Amazon Web Services usage, regardless of your purchase option.

newSavingsPlansCoverageData :: SavingsPlansCoverageData

Create a value of SavingsPlansCoverageData with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:onDemandCost:SavingsPlansCoverageData', savingsPlansCoverageData_onDemandCost - The cost of your Amazon Web Services usage at the public On-Demand rate.

$sel:spendCoveredBySavingsPlans:SavingsPlansCoverageData', savingsPlansCoverageData_spendCoveredBySavingsPlans - The amount of your Amazon Web Services usage that is covered by a Savings Plans.

$sel:coveragePercentage:SavingsPlansCoverageData', savingsPlansCoverageData_coveragePercentage - The percentage of your existing Savings Plans covered usage, divided by all of your eligible Savings Plans usage in an account (or set of accounts).

$sel:totalCost:SavingsPlansCoverageData', savingsPlansCoverageData_totalCost - The total cost of your Amazon Web Services usage, regardless of your purchase option.

savingsPlansCoverageData_onDemandCost :: Lens' SavingsPlansCoverageData (Maybe Text)

The cost of your Amazon Web Services usage at the public On-Demand rate.

savingsPlansCoverageData_spendCoveredBySavingsPlans :: Lens' SavingsPlansCoverageData (Maybe Text)

The amount of your Amazon Web Services usage that is covered by a Savings Plans.

savingsPlansCoverageData_coveragePercentage :: Lens' SavingsPlansCoverageData (Maybe Text)

The percentage of your existing Savings Plans covered usage, divided by all of your eligible Savings Plans usage in an account (or set of accounts).

savingsPlansCoverageData_totalCost :: Lens' SavingsPlansCoverageData (Maybe Text)

The total cost of your Amazon Web Services usage, regardless of your purchase option.