Amazonka.CostExplorer.Types.SavingsPlansPurchaseRecommendationDetail

Description

Documentation

data SavingsPlansPurchaseRecommendationDetail

Details for your recommended Savings Plans.

See: newSavingsPlansPurchaseRecommendationDetail smart constructor.

Constructors

SavingsPlansPurchaseRecommendationDetail'

Fields

currencyCode :: Maybe Text

The currency code that Amazon Web Services used to generate the recommendations and present potential savings.

currentAverageHourlyOnDemandSpend :: Maybe Text

The average value of hourly On-Demand spend over the lookback period of the applicable usage type.

savingsPlansDetails :: Maybe SavingsPlansDetails

Details for your recommended Savings Plans.

currentMinimumHourlyOnDemandSpend :: Maybe Text

The lowest value of hourly On-Demand spend over the lookback period of the applicable usage type.

estimatedROI :: Maybe Text

The estimated return on investment that's based on the recommended Savings Plans that you purchased. This is calculated as estimatedSavingsAmount/ estimatedSPCost*100.

currentMaximumHourlyOnDemandSpend :: Maybe Text

The highest value of hourly On-Demand spend over the lookback period of the applicable usage type.

estimatedSavingsAmount :: Maybe Text

The estimated savings amount that's based on the recommended Savings Plans over the length of the lookback period.

accountId :: Maybe Text

The AccountID the recommendation is generated for.

estimatedMonthlySavingsAmount :: Maybe Text

The estimated monthly savings amount based on the recommended Savings Plans.

estimatedOnDemandCost :: Maybe Text

The remaining On-Demand cost estimated to not be covered by the recommended Savings Plans, over the length of the lookback period.

estimatedOnDemandCostWithCurrentCommitment :: Maybe Text

The estimated On-Demand costs you would expect with no additional commitment, based on your usage of the selected time period and the Savings Plans you own.

upfrontCost :: Maybe Text

The upfront cost of the recommended Savings Plans, based on the selected payment option.

estimatedSPCost :: Maybe Text

The cost of the recommended Savings Plans over the length of the lookback period.

estimatedSavingsPercentage :: Maybe Text

The estimated savings percentage relative to the total cost of applicable On-Demand usage over the lookback period.

estimatedAverageUtilization :: Maybe Text

The estimated utilization of the recommended Savings Plans.

hourlyCommitmentToPurchase :: Maybe Text

The recommended hourly commitment level for the Savings Plans type and the configuration that's based on the usage during the lookback period.

newSavingsPlansPurchaseRecommendationDetail :: SavingsPlansPurchaseRecommendationDetail

Create a value of SavingsPlansPurchaseRecommendationDetail with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:currencyCode:SavingsPlansPurchaseRecommendationDetail', savingsPlansPurchaseRecommendationDetail_currencyCode - The currency code that Amazon Web Services used to generate the recommendations and present potential savings.

$sel:currentAverageHourlyOnDemandSpend:SavingsPlansPurchaseRecommendationDetail', savingsPlansPurchaseRecommendationDetail_currentAverageHourlyOnDemandSpend - The average value of hourly On-Demand spend over the lookback period of the applicable usage type.

$sel:savingsPlansDetails:SavingsPlansPurchaseRecommendationDetail', savingsPlansPurchaseRecommendationDetail_savingsPlansDetails - Details for your recommended Savings Plans.

$sel:currentMinimumHourlyOnDemandSpend:SavingsPlansPurchaseRecommendationDetail', savingsPlansPurchaseRecommendationDetail_currentMinimumHourlyOnDemandSpend - The lowest value of hourly On-Demand spend over the lookback period of the applicable usage type.

$sel:estimatedROI:SavingsPlansPurchaseRecommendationDetail', savingsPlansPurchaseRecommendationDetail_estimatedROI - The estimated return on investment that's based on the recommended Savings Plans that you purchased. This is calculated as estimatedSavingsAmount/ estimatedSPCost*100.

$sel:currentMaximumHourlyOnDemandSpend:SavingsPlansPurchaseRecommendationDetail', savingsPlansPurchaseRecommendationDetail_currentMaximumHourlyOnDemandSpend - The highest value of hourly On-Demand spend over the lookback period of the applicable usage type.

$sel:estimatedSavingsAmount:SavingsPlansPurchaseRecommendationDetail', savingsPlansPurchaseRecommendationDetail_estimatedSavingsAmount - The estimated savings amount that's based on the recommended Savings Plans over the length of the lookback period.

$sel:accountId:SavingsPlansPurchaseRecommendationDetail', savingsPlansPurchaseRecommendationDetail_accountId - The AccountID the recommendation is generated for.

$sel:estimatedMonthlySavingsAmount:SavingsPlansPurchaseRecommendationDetail', savingsPlansPurchaseRecommendationDetail_estimatedMonthlySavingsAmount - The estimated monthly savings amount based on the recommended Savings Plans.

$sel:estimatedOnDemandCost:SavingsPlansPurchaseRecommendationDetail', savingsPlansPurchaseRecommendationDetail_estimatedOnDemandCost - The remaining On-Demand cost estimated to not be covered by the recommended Savings Plans, over the length of the lookback period.

$sel:estimatedOnDemandCostWithCurrentCommitment:SavingsPlansPurchaseRecommendationDetail', savingsPlansPurchaseRecommendationDetail_estimatedOnDemandCostWithCurrentCommitment - The estimated On-Demand costs you would expect with no additional commitment, based on your usage of the selected time period and the Savings Plans you own.

$sel:upfrontCost:SavingsPlansPurchaseRecommendationDetail', savingsPlansPurchaseRecommendationDetail_upfrontCost - The upfront cost of the recommended Savings Plans, based on the selected payment option.

$sel:estimatedSPCost:SavingsPlansPurchaseRecommendationDetail', savingsPlansPurchaseRecommendationDetail_estimatedSPCost - The cost of the recommended Savings Plans over the length of the lookback period.

$sel:estimatedSavingsPercentage:SavingsPlansPurchaseRecommendationDetail', savingsPlansPurchaseRecommendationDetail_estimatedSavingsPercentage - The estimated savings percentage relative to the total cost of applicable On-Demand usage over the lookback period.

$sel:estimatedAverageUtilization:SavingsPlansPurchaseRecommendationDetail', savingsPlansPurchaseRecommendationDetail_estimatedAverageUtilization - The estimated utilization of the recommended Savings Plans.

$sel:hourlyCommitmentToPurchase:SavingsPlansPurchaseRecommendationDetail', savingsPlansPurchaseRecommendationDetail_hourlyCommitmentToPurchase - The recommended hourly commitment level for the Savings Plans type and the configuration that's based on the usage during the lookback period.

savingsPlansPurchaseRecommendationDetail_currencyCode :: Lens' SavingsPlansPurchaseRecommendationDetail (Maybe Text)

The currency code that Amazon Web Services used to generate the recommendations and present potential savings.

savingsPlansPurchaseRecommendationDetail_currentAverageHourlyOnDemandSpend :: Lens' SavingsPlansPurchaseRecommendationDetail (Maybe Text)

The average value of hourly On-Demand spend over the lookback period of the applicable usage type.

savingsPlansPurchaseRecommendationDetail_savingsPlansDetails :: Lens' SavingsPlansPurchaseRecommendationDetail (Maybe SavingsPlansDetails)

Details for your recommended Savings Plans.

savingsPlansPurchaseRecommendationDetail_currentMinimumHourlyOnDemandSpend :: Lens' SavingsPlansPurchaseRecommendationDetail (Maybe Text)

The lowest value of hourly On-Demand spend over the lookback period of the applicable usage type.

savingsPlansPurchaseRecommendationDetail_estimatedROI :: Lens' SavingsPlansPurchaseRecommendationDetail (Maybe Text)

The estimated return on investment that's based on the recommended Savings Plans that you purchased. This is calculated as estimatedSavingsAmount/ estimatedSPCost*100.

savingsPlansPurchaseRecommendationDetail_currentMaximumHourlyOnDemandSpend :: Lens' SavingsPlansPurchaseRecommendationDetail (Maybe Text)

The highest value of hourly On-Demand spend over the lookback period of the applicable usage type.

savingsPlansPurchaseRecommendationDetail_estimatedSavingsAmount :: Lens' SavingsPlansPurchaseRecommendationDetail (Maybe Text)

The estimated savings amount that's based on the recommended Savings Plans over the length of the lookback period.

savingsPlansPurchaseRecommendationDetail_accountId :: Lens' SavingsPlansPurchaseRecommendationDetail (Maybe Text)

The AccountID the recommendation is generated for.

savingsPlansPurchaseRecommendationDetail_estimatedMonthlySavingsAmount :: Lens' SavingsPlansPurchaseRecommendationDetail (Maybe Text)

The estimated monthly savings amount based on the recommended Savings Plans.

savingsPlansPurchaseRecommendationDetail_estimatedOnDemandCost :: Lens' SavingsPlansPurchaseRecommendationDetail (Maybe Text)

The remaining On-Demand cost estimated to not be covered by the recommended Savings Plans, over the length of the lookback period.

savingsPlansPurchaseRecommendationDetail_estimatedOnDemandCostWithCurrentCommitment :: Lens' SavingsPlansPurchaseRecommendationDetail (Maybe Text)

The estimated On-Demand costs you would expect with no additional commitment, based on your usage of the selected time period and the Savings Plans you own.

savingsPlansPurchaseRecommendationDetail_upfrontCost :: Lens' SavingsPlansPurchaseRecommendationDetail (Maybe Text)

The upfront cost of the recommended Savings Plans, based on the selected payment option.

savingsPlansPurchaseRecommendationDetail_estimatedSPCost :: Lens' SavingsPlansPurchaseRecommendationDetail (Maybe Text)

The cost of the recommended Savings Plans over the length of the lookback period.

savingsPlansPurchaseRecommendationDetail_estimatedSavingsPercentage :: Lens' SavingsPlansPurchaseRecommendationDetail (Maybe Text)

The estimated savings percentage relative to the total cost of applicable On-Demand usage over the lookback period.

savingsPlansPurchaseRecommendationDetail_estimatedAverageUtilization :: Lens' SavingsPlansPurchaseRecommendationDetail (Maybe Text)

The estimated utilization of the recommended Savings Plans.

savingsPlansPurchaseRecommendationDetail_hourlyCommitmentToPurchase :: Lens' SavingsPlansPurchaseRecommendationDetail (Maybe Text)

The recommended hourly commitment level for the Savings Plans type and the configuration that's based on the usage during the lookback period.