Amazonka.CostExplorer.Types.SavingsPlansPurchaseRecommendationMetadata

Description

Documentation

data SavingsPlansPurchaseRecommendationMetadata

Metadata about your Savings Plans Purchase Recommendations.

See: newSavingsPlansPurchaseRecommendationMetadata smart constructor.

Constructors

SavingsPlansPurchaseRecommendationMetadata'

Fields

recommendationId :: Maybe Text

The unique identifier for the recommendation set.

generationTimestamp :: Maybe Text

The timestamp showing when the recommendations were generated.

additionalMetadata :: Maybe Text

Additional metadata that might be applicable to the recommendation.

newSavingsPlansPurchaseRecommendationMetadata :: SavingsPlansPurchaseRecommendationMetadata

Create a value of SavingsPlansPurchaseRecommendationMetadata with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:recommendationId:SavingsPlansPurchaseRecommendationMetadata', savingsPlansPurchaseRecommendationMetadata_recommendationId - The unique identifier for the recommendation set.

$sel:generationTimestamp:SavingsPlansPurchaseRecommendationMetadata', savingsPlansPurchaseRecommendationMetadata_generationTimestamp - The timestamp showing when the recommendations were generated.

$sel:additionalMetadata:SavingsPlansPurchaseRecommendationMetadata', savingsPlansPurchaseRecommendationMetadata_additionalMetadata - Additional metadata that might be applicable to the recommendation.

savingsPlansPurchaseRecommendationMetadata_recommendationId :: Lens' SavingsPlansPurchaseRecommendationMetadata (Maybe Text)

The unique identifier for the recommendation set.

savingsPlansPurchaseRecommendationMetadata_generationTimestamp :: Lens' SavingsPlansPurchaseRecommendationMetadata (Maybe Text)

The timestamp showing when the recommendations were generated.

savingsPlansPurchaseRecommendationMetadata_additionalMetadata :: Lens' SavingsPlansPurchaseRecommendationMetadata (Maybe Text)

Additional metadata that might be applicable to the recommendation.