Amazonka.CostExplorer.Types.SavingsPlansUtilizationAggregates

Description

Documentation

data SavingsPlansUtilizationAggregates

The aggregated utilization metrics for your Savings Plans usage.

See: newSavingsPlansUtilizationAggregates smart constructor.

Constructors

SavingsPlansUtilizationAggregates'

Fields

amortizedCommitment :: Maybe SavingsPlansAmortizedCommitment

The total amortized commitment for a Savings Plans. This includes the sum of the upfront and recurring Savings Plans fees.

savings :: Maybe SavingsPlansSavings

The amount saved by using existing Savings Plans. Savings returns both net savings from Savings Plans, as well as the onDemandCostEquivalent of the Savings Plans when considering the utilization rate.

utilization :: SavingsPlansUtilization

A ratio of your effectiveness of using existing Savings Plans to apply to workloads that are Savings Plans eligible.

newSavingsPlansUtilizationAggregates

Arguments

:: SavingsPlansUtilization
$sel:utilization:SavingsPlansUtilizationAggregates'

-> SavingsPlansUtilizationAggregates

Create a value of SavingsPlansUtilizationAggregates with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:amortizedCommitment:SavingsPlansUtilizationAggregates', savingsPlansUtilizationAggregates_amortizedCommitment - The total amortized commitment for a Savings Plans. This includes the sum of the upfront and recurring Savings Plans fees.

$sel:savings:SavingsPlansUtilizationAggregates', savingsPlansUtilizationAggregates_savings - The amount saved by using existing Savings Plans. Savings returns both net savings from Savings Plans, as well as the onDemandCostEquivalent of the Savings Plans when considering the utilization rate.

$sel:utilization:SavingsPlansUtilizationAggregates', savingsPlansUtilizationAggregates_utilization - A ratio of your effectiveness of using existing Savings Plans to apply to workloads that are Savings Plans eligible.

savingsPlansUtilizationAggregates_amortizedCommitment :: Lens' SavingsPlansUtilizationAggregates (Maybe SavingsPlansAmortizedCommitment)

The total amortized commitment for a Savings Plans. This includes the sum of the upfront and recurring Savings Plans fees.

savingsPlansUtilizationAggregates_savings :: Lens' SavingsPlansUtilizationAggregates (Maybe SavingsPlansSavings)

The amount saved by using existing Savings Plans. Savings returns both net savings from Savings Plans, as well as the onDemandCostEquivalent of the Savings Plans when considering the utilization rate.

savingsPlansUtilizationAggregates_utilization :: Lens' SavingsPlansUtilizationAggregates SavingsPlansUtilization

A ratio of your effectiveness of using existing Savings Plans to apply to workloads that are Savings Plans eligible.