Amazonka.CostExplorer.Types.ServiceSpecification

Description

Documentation

data ServiceSpecification

Hardware specifications for the service that you want recommendations for.

See: newServiceSpecification smart constructor.

Constructors

ServiceSpecification'

Fields

eC2Specification :: Maybe EC2Specification

The Amazon EC2 hardware specifications that you want Amazon Web Services to provide recommendations for.

newServiceSpecification :: ServiceSpecification

Create a value of ServiceSpecification with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:eC2Specification:ServiceSpecification', serviceSpecification_eC2Specification - The Amazon EC2 hardware specifications that you want Amazon Web Services to provide recommendations for.

serviceSpecification_eC2Specification :: Lens' ServiceSpecification (Maybe EC2Specification)

The Amazon EC2 hardware specifications that you want Amazon Web Services to provide recommendations for.