Amazonka.CustomerProfiles.CreateDomain

Description

Creates a domain, which is a container for all customer data, such as customer profile attributes, object types, profile keys, and encryption keys. You can create multiple domains, and each domain can have multiple third-party integrations.

Each Amazon Connect instance can be associated with only one domain. Multiple Amazon Connect instances can be associated with one domain.

Use this API or UpdateDomain to enable identity resolution: set Matching to true.

Creating a Request

data CreateDomain

See: newCreateDomain smart constructor.

Constructors

CreateDomain'

Fields

defaultEncryptionKey :: Maybe Text

The default encryption key, which is an AWS managed key, is used when no specific type of encryption key is specified. It is used to encrypt all data before it is placed in permanent or semi-permanent storage.

matching :: Maybe MatchingRequest

The process of matching duplicate profiles. If Matching = true, Amazon Connect Customer Profiles starts a weekly batch process every Saturday at 12AM UTC to detect duplicate profiles in your domains. After that batch process completes, use the GetMatches API to return and review the results.

deadLetterQueueUrl :: Maybe Text

The URL of the SQS dead letter queue, which is used for reporting errors associated with ingesting data from third party applications. You must set up a policy on the DeadLetterQueue for the SendMessage operation to enable Amazon Connect Customer Profiles to send messages to the DeadLetterQueue.

tags :: Maybe (HashMap Text Text)

The tags used to organize, track, or control access for this resource.

domainName :: Text

The unique name of the domain.

defaultExpirationDays :: Natural

The default number of days until the data within the domain expires.

newCreateDomain

Arguments

:: Text
$sel:domainName:CreateDomain'

-> Natural
$sel:defaultExpirationDays:CreateDomain'

-> CreateDomain

Create a value of CreateDomain with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:defaultEncryptionKey:CreateDomain', createDomain_defaultEncryptionKey - The default encryption key, which is an AWS managed key, is used when no specific type of encryption key is specified. It is used to encrypt all data before it is placed in permanent or semi-permanent storage.

$sel:matching:CreateDomain', createDomain_matching - The process of matching duplicate profiles. If Matching = true, Amazon Connect Customer Profiles starts a weekly batch process every Saturday at 12AM UTC to detect duplicate profiles in your domains. After that batch process completes, use the GetMatches API to return and review the results.

$sel:deadLetterQueueUrl:CreateDomain', createDomain_deadLetterQueueUrl - The URL of the SQS dead letter queue, which is used for reporting errors associated with ingesting data from third party applications. You must set up a policy on the DeadLetterQueue for the SendMessage operation to enable Amazon Connect Customer Profiles to send messages to the DeadLetterQueue.

$sel:tags:CreateDomain', createDomain_tags - The tags used to organize, track, or control access for this resource.

$sel:domainName:CreateDomain', createDomain_domainName - The unique name of the domain.

$sel:defaultExpirationDays:CreateDomain', createDomain_defaultExpirationDays - The default number of days until the data within the domain expires.

Request Lenses

createDomain_defaultEncryptionKey :: Lens' CreateDomain (Maybe Text)

The default encryption key, which is an AWS managed key, is used when no specific type of encryption key is specified. It is used to encrypt all data before it is placed in permanent or semi-permanent storage.

createDomain_matching :: Lens' CreateDomain (Maybe MatchingRequest)

The process of matching duplicate profiles. If Matching = true, Amazon Connect Customer Profiles starts a weekly batch process every Saturday at 12AM UTC to detect duplicate profiles in your domains. After that batch process completes, use the GetMatches API to return and review the results.

createDomain_deadLetterQueueUrl :: Lens' CreateDomain (Maybe Text)

The URL of the SQS dead letter queue, which is used for reporting errors associated with ingesting data from third party applications. You must set up a policy on the DeadLetterQueue for the SendMessage operation to enable Amazon Connect Customer Profiles to send messages to the DeadLetterQueue.

createDomain_tags :: Lens' CreateDomain (Maybe (HashMap Text Text))

The tags used to organize, track, or control access for this resource.

createDomain_domainName :: Lens' CreateDomain Text

The unique name of the domain.

createDomain_defaultExpirationDays :: Lens' CreateDomain Natural

The default number of days until the data within the domain expires.

Destructuring the Response

data CreateDomainResponse

See: newCreateDomainResponse smart constructor.

Constructors

CreateDomainResponse'

Fields

defaultEncryptionKey :: Maybe Text

The default encryption key, which is an AWS managed key, is used when no specific type of encryption key is specified. It is used to encrypt all data before it is placed in permanent or semi-permanent storage.

matching :: Maybe MatchingResponse

The process of matching duplicate profiles. If Matching = true, Amazon Connect Customer Profiles starts a weekly batch process every Saturday at 12AM UTC to detect duplicate profiles in your domains. After that batch process completes, use the GetMatches API to return and review the results.

deadLetterQueueUrl :: Maybe Text

The URL of the SQS dead letter queue, which is used for reporting errors associated with ingesting data from third party applications.

tags :: Maybe (HashMap Text Text)

The tags used to organize, track, or control access for this resource.

httpStatus :: Int

The response's http status code.

domainName :: Text

The unique name of the domain.

defaultExpirationDays :: Natural

The default number of days until the data within the domain expires.

createdAt :: POSIX

The timestamp of when the domain was created.

lastUpdatedAt :: POSIX

The timestamp of when the domain was most recently edited.

newCreateDomainResponse

Arguments

:: Int
$sel:httpStatus:CreateDomainResponse'

-> Text
$sel:domainName:CreateDomain'

-> Natural
$sel:defaultExpirationDays:CreateDomain'

-> UTCTime
$sel:createdAt:CreateDomainResponse'

-> UTCTime
$sel:lastUpdatedAt:CreateDomainResponse'

-> CreateDomainResponse

Create a value of CreateDomainResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:defaultEncryptionKey:CreateDomain', createDomainResponse_defaultEncryptionKey - The default encryption key, which is an AWS managed key, is used when no specific type of encryption key is specified. It is used to encrypt all data before it is placed in permanent or semi-permanent storage.

$sel:matching:CreateDomain', createDomainResponse_matching - The process of matching duplicate profiles. If Matching = true, Amazon Connect Customer Profiles starts a weekly batch process every Saturday at 12AM UTC to detect duplicate profiles in your domains. After that batch process completes, use the GetMatches API to return and review the results.

$sel:deadLetterQueueUrl:CreateDomain', createDomainResponse_deadLetterQueueUrl - The URL of the SQS dead letter queue, which is used for reporting errors associated with ingesting data from third party applications.

$sel:tags:CreateDomain', createDomainResponse_tags - The tags used to organize, track, or control access for this resource.

$sel:httpStatus:CreateDomainResponse', createDomainResponse_httpStatus - The response's http status code.

$sel:domainName:CreateDomain', createDomainResponse_domainName - The unique name of the domain.

$sel:defaultExpirationDays:CreateDomain', createDomainResponse_defaultExpirationDays - The default number of days until the data within the domain expires.

$sel:createdAt:CreateDomainResponse', createDomainResponse_createdAt - The timestamp of when the domain was created.

$sel:lastUpdatedAt:CreateDomainResponse', createDomainResponse_lastUpdatedAt - The timestamp of when the domain was most recently edited.

Response Lenses

createDomainResponse_defaultEncryptionKey :: Lens' CreateDomainResponse (Maybe Text)

The default encryption key, which is an AWS managed key, is used when no specific type of encryption key is specified. It is used to encrypt all data before it is placed in permanent or semi-permanent storage.

createDomainResponse_matching :: Lens' CreateDomainResponse (Maybe MatchingResponse)

The process of matching duplicate profiles. If Matching = true, Amazon Connect Customer Profiles starts a weekly batch process every Saturday at 12AM UTC to detect duplicate profiles in your domains. After that batch process completes, use the GetMatches API to return and review the results.

createDomainResponse_deadLetterQueueUrl :: Lens' CreateDomainResponse (Maybe Text)

The URL of the SQS dead letter queue, which is used for reporting errors associated with ingesting data from third party applications.

createDomainResponse_tags :: Lens' CreateDomainResponse (Maybe (HashMap Text Text))

The tags used to organize, track, or control access for this resource.

createDomainResponse_httpStatus :: Lens' CreateDomainResponse Int

The response's http status code.

createDomainResponse_domainName :: Lens' CreateDomainResponse Text

The unique name of the domain.

createDomainResponse_defaultExpirationDays :: Lens' CreateDomainResponse Natural

The default number of days until the data within the domain expires.

createDomainResponse_createdAt :: Lens' CreateDomainResponse UTCTime

The timestamp of when the domain was created.

createDomainResponse_lastUpdatedAt :: Lens' CreateDomainResponse UTCTime

The timestamp of when the domain was most recently edited.