Amazonka.CustomerProfiles.CreateProfile

Description

Creates a standard profile.

A standard profile represents the following attributes for a customer profile in a domain.

Creating a Request

data CreateProfile

See: newCreateProfile smart constructor.

Constructors

CreateProfile'

Fields

shippingAddress :: Maybe Address

The customer’s shipping address.

mobilePhoneNumber :: Maybe Text

The customer’s mobile phone number.

mailingAddress :: Maybe Address

The customer’s mailing address.

middleName :: Maybe Text

The customer’s middle name.

personalEmailAddress :: Maybe Text

The customer’s personal email address.

lastName :: Maybe Text

The customer’s last name.

additionalInformation :: Maybe Text

Any additional information relevant to the customer’s profile.

homePhoneNumber :: Maybe Text

The customer’s home phone number.

address :: Maybe Address

A generic address associated with the customer that is not mailing, shipping, or billing.

partyType :: Maybe PartyType

The type of profile used to describe the customer.

businessEmailAddress :: Maybe Text

The customer’s business email address.

attributes :: Maybe (HashMap Text Text)

A key value pair of attributes of a customer profile.

gender :: Maybe Gender

The gender with which the customer identifies.

phoneNumber :: Maybe Text

The customer’s phone number, which has not been specified as a mobile, home, or business number.

accountNumber :: Maybe Text

A unique account number that you have given to the customer.

emailAddress :: Maybe Text

The customer’s email address, which has not been specified as a personal or business address.

firstName :: Maybe Text

The customer’s first name.

billingAddress :: Maybe Address

The customer’s billing address.

businessPhoneNumber :: Maybe Text

The customer’s business phone number.

birthDate :: Maybe Text

The customer’s birth date.

businessName :: Maybe Text

The name of the customer’s business.

domainName :: Text

The unique name of the domain.

newCreateProfile

Arguments

:: Text
$sel:domainName:CreateProfile'

-> CreateProfile

Create a value of CreateProfile with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:shippingAddress:CreateProfile', createProfile_shippingAddress - The customer’s shipping address.

$sel:mobilePhoneNumber:CreateProfile', createProfile_mobilePhoneNumber - The customer’s mobile phone number.

$sel:mailingAddress:CreateProfile', createProfile_mailingAddress - The customer’s mailing address.

$sel:middleName:CreateProfile', createProfile_middleName - The customer’s middle name.

$sel:personalEmailAddress:CreateProfile', createProfile_personalEmailAddress - The customer’s personal email address.

$sel:lastName:CreateProfile', createProfile_lastName - The customer’s last name.

$sel:additionalInformation:CreateProfile', createProfile_additionalInformation - Any additional information relevant to the customer’s profile.

$sel:homePhoneNumber:CreateProfile', createProfile_homePhoneNumber - The customer’s home phone number.

$sel:address:CreateProfile', createProfile_address - A generic address associated with the customer that is not mailing, shipping, or billing.

$sel:partyType:CreateProfile', createProfile_partyType - The type of profile used to describe the customer.

$sel:businessEmailAddress:CreateProfile', createProfile_businessEmailAddress - The customer’s business email address.

$sel:attributes:CreateProfile', createProfile_attributes - A key value pair of attributes of a customer profile.

$sel:gender:CreateProfile', createProfile_gender - The gender with which the customer identifies.

$sel:phoneNumber:CreateProfile', createProfile_phoneNumber - The customer’s phone number, which has not been specified as a mobile, home, or business number.

$sel:accountNumber:CreateProfile', createProfile_accountNumber - A unique account number that you have given to the customer.

$sel:emailAddress:CreateProfile', createProfile_emailAddress - The customer’s email address, which has not been specified as a personal or business address.

$sel:firstName:CreateProfile', createProfile_firstName - The customer’s first name.

$sel:billingAddress:CreateProfile', createProfile_billingAddress - The customer’s billing address.

$sel:businessPhoneNumber:CreateProfile', createProfile_businessPhoneNumber - The customer’s business phone number.

$sel:birthDate:CreateProfile', createProfile_birthDate - The customer’s birth date.

$sel:businessName:CreateProfile', createProfile_businessName - The name of the customer’s business.

$sel:domainName:CreateProfile', createProfile_domainName - The unique name of the domain.

Request Lenses

createProfile_shippingAddress :: Lens' CreateProfile (Maybe Address)

The customer’s shipping address.

createProfile_mobilePhoneNumber :: Lens' CreateProfile (Maybe Text)

The customer’s mobile phone number.

createProfile_mailingAddress :: Lens' CreateProfile (Maybe Address)

The customer’s mailing address.

createProfile_middleName :: Lens' CreateProfile (Maybe Text)

The customer’s middle name.

createProfile_personalEmailAddress :: Lens' CreateProfile (Maybe Text)

The customer’s personal email address.

createProfile_lastName :: Lens' CreateProfile (Maybe Text)

The customer’s last name.

createProfile_additionalInformation :: Lens' CreateProfile (Maybe Text)

Any additional information relevant to the customer’s profile.

createProfile_homePhoneNumber :: Lens' CreateProfile (Maybe Text)

The customer’s home phone number.

createProfile_address :: Lens' CreateProfile (Maybe Address)

A generic address associated with the customer that is not mailing, shipping, or billing.

createProfile_partyType :: Lens' CreateProfile (Maybe PartyType)

The type of profile used to describe the customer.

createProfile_businessEmailAddress :: Lens' CreateProfile (Maybe Text)

The customer’s business email address.

createProfile_attributes :: Lens' CreateProfile (Maybe (HashMap Text Text))

A key value pair of attributes of a customer profile.

createProfile_gender :: Lens' CreateProfile (Maybe Gender)

The gender with which the customer identifies.

createProfile_phoneNumber :: Lens' CreateProfile (Maybe Text)

The customer’s phone number, which has not been specified as a mobile, home, or business number.

createProfile_accountNumber :: Lens' CreateProfile (Maybe Text)

A unique account number that you have given to the customer.

createProfile_emailAddress :: Lens' CreateProfile (Maybe Text)

The customer’s email address, which has not been specified as a personal or business address.

createProfile_firstName :: Lens' CreateProfile (Maybe Text)

The customer’s first name.

createProfile_billingAddress :: Lens' CreateProfile (Maybe Address)

The customer’s billing address.

createProfile_businessPhoneNumber :: Lens' CreateProfile (Maybe Text)

The customer’s business phone number.

createProfile_birthDate :: Lens' CreateProfile (Maybe Text)

The customer’s birth date.

createProfile_businessName :: Lens' CreateProfile (Maybe Text)

The name of the customer’s business.

createProfile_domainName :: Lens' CreateProfile Text

The unique name of the domain.

Destructuring the Response

data CreateProfileResponse

See: newCreateProfileResponse smart constructor.

Constructors

CreateProfileResponse'

Fields

httpStatus :: Int

The response's http status code.

profileId :: Text

The unique identifier of a customer profile.

newCreateProfileResponse

Arguments

:: Int
$sel:httpStatus:CreateProfileResponse'

-> Text
$sel:profileId:CreateProfileResponse'

-> CreateProfileResponse

Create a value of CreateProfileResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:httpStatus:CreateProfileResponse', createProfileResponse_httpStatus - The response's http status code.

$sel:profileId:CreateProfileResponse', createProfileResponse_profileId - The unique identifier of a customer profile.

Response Lenses

createProfileResponse_httpStatus :: Lens' CreateProfileResponse Int

The response's http status code.

createProfileResponse_profileId :: Lens' CreateProfileResponse Text

The unique identifier of a customer profile.