Amazonka.CustomerProfiles.GetIntegration

Description

Returns an integration for a domain.

Creating a Request

data GetIntegration

See: newGetIntegration smart constructor.

Constructors

GetIntegration'

Fields

domainName :: Text

The unique name of the domain.

uri :: Text

The URI of the S3 bucket or any other type of data source.

newGetIntegration

Arguments

:: Text
$sel:domainName:GetIntegration'

-> Text
$sel:uri:GetIntegration'

-> GetIntegration

Create a value of GetIntegration with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:domainName:GetIntegration', getIntegration_domainName - The unique name of the domain.

$sel:uri:GetIntegration', getIntegration_uri - The URI of the S3 bucket or any other type of data source.

Request Lenses

getIntegration_domainName :: Lens' GetIntegration Text

The unique name of the domain.

getIntegration_uri :: Lens' GetIntegration Text

The URI of the S3 bucket or any other type of data source.

Destructuring the Response

data GetIntegrationResponse

See: newGetIntegrationResponse smart constructor.

Constructors

GetIntegrationResponse'

Fields

tags :: Maybe (HashMap Text Text)

The tags used to organize, track, or control access for this resource.

httpStatus :: Int

The response's http status code.

domainName :: Text

The unique name of the domain.

uri :: Text

The URI of the S3 bucket or any other type of data source.

objectTypeName :: Text

The name of the profile object type.

createdAt :: POSIX

The timestamp of when the domain was created.

lastUpdatedAt :: POSIX

The timestamp of when the domain was most recently edited.

newGetIntegrationResponse

Arguments

:: Int
$sel:httpStatus:GetIntegrationResponse'

-> Text
$sel:domainName:GetIntegration'

-> Text
$sel:uri:GetIntegration'

-> Text
$sel:objectTypeName:GetIntegrationResponse'

-> UTCTime
$sel:createdAt:GetIntegrationResponse'

-> UTCTime
$sel:lastUpdatedAt:GetIntegrationResponse'

-> GetIntegrationResponse

Create a value of GetIntegrationResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:tags:GetIntegrationResponse', getIntegrationResponse_tags - The tags used to organize, track, or control access for this resource.

$sel:httpStatus:GetIntegrationResponse', getIntegrationResponse_httpStatus - The response's http status code.

$sel:domainName:GetIntegration', getIntegrationResponse_domainName - The unique name of the domain.

$sel:uri:GetIntegration', getIntegrationResponse_uri - The URI of the S3 bucket or any other type of data source.

$sel:objectTypeName:GetIntegrationResponse', getIntegrationResponse_objectTypeName - The name of the profile object type.

$sel:createdAt:GetIntegrationResponse', getIntegrationResponse_createdAt - The timestamp of when the domain was created.

$sel:lastUpdatedAt:GetIntegrationResponse', getIntegrationResponse_lastUpdatedAt - The timestamp of when the domain was most recently edited.

Response Lenses

getIntegrationResponse_tags :: Lens' GetIntegrationResponse (Maybe (HashMap Text Text))

The tags used to organize, track, or control access for this resource.

getIntegrationResponse_httpStatus :: Lens' GetIntegrationResponse Int

The response's http status code.

getIntegrationResponse_domainName :: Lens' GetIntegrationResponse Text

The unique name of the domain.

getIntegrationResponse_uri :: Lens' GetIntegrationResponse Text

The URI of the S3 bucket or any other type of data source.

getIntegrationResponse_objectTypeName :: Lens' GetIntegrationResponse Text

The name of the profile object type.

getIntegrationResponse_createdAt :: Lens' GetIntegrationResponse UTCTime

The timestamp of when the domain was created.

getIntegrationResponse_lastUpdatedAt :: Lens' GetIntegrationResponse UTCTime

The timestamp of when the domain was most recently edited.