Amazonka.CustomerProfiles.GetMatches

Description

This API is in preview release for Amazon Connect and subject to change.

Before calling this API, use CreateDomain or UpdateDomain to enable identity resolution: set Matching to true.

GetMatches returns potentially matching profiles, based on the results of the latest run of a machine learning process.

Amazon Connect starts a batch process every Saturday at 12AM UTC to identify matching profiles. The results are returned up to seven days after the Saturday run.

Amazon Connect uses the following profile attributes to identify matches:

PhoneNumber

HomePhoneNumber

BusinessPhoneNumber

MobilePhoneNumber

EmailAddress

PersonalEmailAddress

BusinessEmailAddress

FullName

BusinessName

For example, two or more profiles—with spelling mistakes such as __John Doe and Jhn Doe__, or different casing email addresses such as JOHN_DOE@ANYCOMPANY.COM and johndoe@anycompany.com, or different phone number formats such as 555-010-0000 and +1-555-010-0000—can be detected as belonging to the same customer John Doe and merged into a unified profile.

Creating a Request

data GetMatches

See: newGetMatches smart constructor.

Constructors

GetMatches'

Fields

nextToken :: Maybe Text

The token for the next set of results. Use the value returned in the previous response in the next request to retrieve the next set of results.

maxResults :: Maybe Natural

The maximum number of results to return per page.

domainName :: Text

The unique name of the domain.

newGetMatches

Arguments

:: Text
$sel:domainName:GetMatches'

-> GetMatches

Create a value of GetMatches with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:nextToken:GetMatches', getMatches_nextToken - The token for the next set of results. Use the value returned in the previous response in the next request to retrieve the next set of results.

$sel:maxResults:GetMatches', getMatches_maxResults - The maximum number of results to return per page.

$sel:domainName:GetMatches', getMatches_domainName - The unique name of the domain.

Request Lenses

getMatches_nextToken :: Lens' GetMatches (Maybe Text)

The token for the next set of results. Use the value returned in the previous response in the next request to retrieve the next set of results.

getMatches_maxResults :: Lens' GetMatches (Maybe Natural)

The maximum number of results to return per page.

getMatches_domainName :: Lens' GetMatches Text

The unique name of the domain.

Destructuring the Response

data GetMatchesResponse

See: newGetMatchesResponse smart constructor.

Constructors

GetMatchesResponse'

Fields

matches :: Maybe [MatchItem]

The list of matched profiles for this instance.

matchGenerationDate :: Maybe POSIX

The timestamp this version of Match Result generated.

nextToken :: Maybe Text

If there are additional results, this is the token for the next set of results.

potentialMatches :: Maybe Natural

The number of potential matches found.

httpStatus :: Int

The response's http status code.

newGetMatchesResponse

Arguments

:: Int
$sel:httpStatus:GetMatchesResponse'

-> GetMatchesResponse

Create a value of GetMatchesResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:matches:GetMatchesResponse', getMatchesResponse_matches - The list of matched profiles for this instance.

$sel:matchGenerationDate:GetMatchesResponse', getMatchesResponse_matchGenerationDate - The timestamp this version of Match Result generated.

$sel:nextToken:GetMatches', getMatchesResponse_nextToken - If there are additional results, this is the token for the next set of results.

$sel:potentialMatches:GetMatchesResponse', getMatchesResponse_potentialMatches - The number of potential matches found.

$sel:httpStatus:GetMatchesResponse', getMatchesResponse_httpStatus - The response's http status code.

Response Lenses

getMatchesResponse_matches :: Lens' GetMatchesResponse (Maybe [MatchItem])

The list of matched profiles for this instance.

getMatchesResponse_matchGenerationDate :: Lens' GetMatchesResponse (Maybe UTCTime)

The timestamp this version of Match Result generated.

getMatchesResponse_nextToken :: Lens' GetMatchesResponse (Maybe Text)

If there are additional results, this is the token for the next set of results.

getMatchesResponse_potentialMatches :: Lens' GetMatchesResponse (Maybe Natural)

The number of potential matches found.

getMatchesResponse_httpStatus :: Lens' GetMatchesResponse Int

The response's http status code.