Amazonka.CustomerProfiles.GetProfileObjectType

Description

Returns the object types for a specific domain.

Creating a Request

data GetProfileObjectType

See: newGetProfileObjectType smart constructor.

Constructors

GetProfileObjectType'

Fields

domainName :: Text

The unique name of the domain.

objectTypeName :: Text

The name of the profile object type.

newGetProfileObjectType

Arguments

:: Text
$sel:domainName:GetProfileObjectType'

-> Text
$sel:objectTypeName:GetProfileObjectType'

-> GetProfileObjectType

Create a value of GetProfileObjectType with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:domainName:GetProfileObjectType', getProfileObjectType_domainName - The unique name of the domain.

$sel:objectTypeName:GetProfileObjectType', getProfileObjectType_objectTypeName - The name of the profile object type.

Request Lenses

getProfileObjectType_domainName :: Lens' GetProfileObjectType Text

The unique name of the domain.

getProfileObjectType_objectTypeName :: Lens' GetProfileObjectType Text

The name of the profile object type.

Destructuring the Response

data GetProfileObjectTypeResponse

See: newGetProfileObjectTypeResponse smart constructor.

Constructors

GetProfileObjectTypeResponse'

Fields

expirationDays :: Maybe Natural

The number of days until the data in the object expires.

lastUpdatedAt :: Maybe POSIX

The timestamp of when the domain was most recently edited.

createdAt :: Maybe POSIX

The timestamp of when the domain was created.

templateId :: Maybe Text

A unique identifier for the object template.

keys :: Maybe (HashMap Text [ObjectTypeKey])

A list of unique keys that can be used to map data to the profile.

encryptionKey :: Maybe Text

The customer-provided key to encrypt the profile object that will be created in this profile object type.

allowProfileCreation :: Maybe Bool

Indicates whether a profile should be created when data is received if one doesn’t exist for an object of this type. The default is FALSE. If the AllowProfileCreation flag is set to FALSE, then the service tries to fetch a standard profile and associate this object with the profile. If it is set to TRUE, and if no match is found, then the service creates a new standard profile.

fields :: Maybe (HashMap Text ObjectTypeField)

A map of the name and ObjectType field.

tags :: Maybe (HashMap Text Text)

The tags used to organize, track, or control access for this resource.

httpStatus :: Int

The response's http status code.

objectTypeName :: Text

The name of the profile object type.

description :: Text

The description of the profile object type.

newGetProfileObjectTypeResponse

Arguments

:: Int
$sel:httpStatus:GetProfileObjectTypeResponse'

-> Text
$sel:objectTypeName:GetProfileObjectType'

-> Text
$sel:description:GetProfileObjectTypeResponse'

-> GetProfileObjectTypeResponse

Create a value of GetProfileObjectTypeResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:expirationDays:GetProfileObjectTypeResponse', getProfileObjectTypeResponse_expirationDays - The number of days until the data in the object expires.

$sel:lastUpdatedAt:GetProfileObjectTypeResponse', getProfileObjectTypeResponse_lastUpdatedAt - The timestamp of when the domain was most recently edited.

$sel:createdAt:GetProfileObjectTypeResponse', getProfileObjectTypeResponse_createdAt - The timestamp of when the domain was created.

$sel:templateId:GetProfileObjectTypeResponse', getProfileObjectTypeResponse_templateId - A unique identifier for the object template.

$sel:keys:GetProfileObjectTypeResponse', getProfileObjectTypeResponse_keys - A list of unique keys that can be used to map data to the profile.

$sel:encryptionKey:GetProfileObjectTypeResponse', getProfileObjectTypeResponse_encryptionKey - The customer-provided key to encrypt the profile object that will be created in this profile object type.

$sel:allowProfileCreation:GetProfileObjectTypeResponse', getProfileObjectTypeResponse_allowProfileCreation - Indicates whether a profile should be created when data is received if one doesn’t exist for an object of this type. The default is FALSE. If the AllowProfileCreation flag is set to FALSE, then the service tries to fetch a standard profile and associate this object with the profile. If it is set to TRUE, and if no match is found, then the service creates a new standard profile.

$sel:fields:GetProfileObjectTypeResponse', getProfileObjectTypeResponse_fields - A map of the name and ObjectType field.

$sel:tags:GetProfileObjectTypeResponse', getProfileObjectTypeResponse_tags - The tags used to organize, track, or control access for this resource.

$sel:httpStatus:GetProfileObjectTypeResponse', getProfileObjectTypeResponse_httpStatus - The response's http status code.

$sel:objectTypeName:GetProfileObjectType', getProfileObjectTypeResponse_objectTypeName - The name of the profile object type.

$sel:description:GetProfileObjectTypeResponse', getProfileObjectTypeResponse_description - The description of the profile object type.

Response Lenses

getProfileObjectTypeResponse_expirationDays :: Lens' GetProfileObjectTypeResponse (Maybe Natural)

The number of days until the data in the object expires.

getProfileObjectTypeResponse_lastUpdatedAt :: Lens' GetProfileObjectTypeResponse (Maybe UTCTime)

The timestamp of when the domain was most recently edited.

getProfileObjectTypeResponse_createdAt :: Lens' GetProfileObjectTypeResponse (Maybe UTCTime)

The timestamp of when the domain was created.

getProfileObjectTypeResponse_templateId :: Lens' GetProfileObjectTypeResponse (Maybe Text)

A unique identifier for the object template.

getProfileObjectTypeResponse_keys :: Lens' GetProfileObjectTypeResponse (Maybe (HashMap Text [ObjectTypeKey]))

A list of unique keys that can be used to map data to the profile.

getProfileObjectTypeResponse_encryptionKey :: Lens' GetProfileObjectTypeResponse (Maybe Text)

The customer-provided key to encrypt the profile object that will be created in this profile object type.

getProfileObjectTypeResponse_allowProfileCreation :: Lens' GetProfileObjectTypeResponse (Maybe Bool)

Indicates whether a profile should be created when data is received if one doesn’t exist for an object of this type. The default is FALSE. If the AllowProfileCreation flag is set to FALSE, then the service tries to fetch a standard profile and associate this object with the profile. If it is set to TRUE, and if no match is found, then the service creates a new standard profile.

getProfileObjectTypeResponse_fields :: Lens' GetProfileObjectTypeResponse (Maybe (HashMap Text ObjectTypeField))

A map of the name and ObjectType field.

getProfileObjectTypeResponse_tags :: Lens' GetProfileObjectTypeResponse (Maybe (HashMap Text Text))

The tags used to organize, track, or control access for this resource.

getProfileObjectTypeResponse_httpStatus :: Lens' GetProfileObjectTypeResponse Int

The response's http status code.

getProfileObjectTypeResponse_objectTypeName :: Lens' GetProfileObjectTypeResponse Text

The name of the profile object type.

getProfileObjectTypeResponse_description :: Lens' GetProfileObjectTypeResponse Text

The description of the profile object type.