Amazonka.CustomerProfiles.MergeProfiles

Description

This API is in preview release for Amazon Connect and subject to change.

Runs an AWS Lambda job that does the following:

All the profileKeys in the ProfileToBeMerged will be moved to the main profile.

All the objects in the ProfileToBeMerged will be moved to the main profile.

All the ProfileToBeMerged will be deleted at the end.

All the profileKeys in the ProfileIdsToBeMerged will be moved to the main profile.

Standard fields are merged as follows:

Fields are always "union"-ed if there are no conflicts in standard fields or attributeKeys.

When there are conflicting fields:

If no SourceProfileIds entry is specified, the main Profile value is always taken.

If a SourceProfileIds entry is specified, the specified profileId is always taken, even if it is a NULL value.

You can use MergeProfiles together with GetMatches, which returns potentially matching profiles, or use it with the results of another matching system. After profiles have been merged, they cannot be separated (unmerged).

Creating a Request

data MergeProfiles

See: newMergeProfiles smart constructor.

Constructors

MergeProfiles'

Fields

fieldSourceProfileIds :: Maybe FieldSourceProfileIds

The identifiers of the fields in the profile that has the information you want to apply to the merge. For example, say you want to merge EmailAddress from Profile1 into MainProfile. This would be the identifier of the EmailAddress field in Profile1.

domainName :: Text

The unique name of the domain.

mainProfileId :: Text

The identifier of the profile to be taken.

profileIdsToBeMerged :: NonEmpty Text

The identifier of the profile to be merged into MainProfileId.

newMergeProfiles

Arguments

:: Text
$sel:domainName:MergeProfiles'

-> Text
$sel:mainProfileId:MergeProfiles'

-> NonEmpty Text
$sel:profileIdsToBeMerged:MergeProfiles'

-> MergeProfiles

Create a value of MergeProfiles with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:fieldSourceProfileIds:MergeProfiles', mergeProfiles_fieldSourceProfileIds - The identifiers of the fields in the profile that has the information you want to apply to the merge. For example, say you want to merge EmailAddress from Profile1 into MainProfile. This would be the identifier of the EmailAddress field in Profile1.

$sel:domainName:MergeProfiles', mergeProfiles_domainName - The unique name of the domain.

$sel:mainProfileId:MergeProfiles', mergeProfiles_mainProfileId - The identifier of the profile to be taken.

$sel:profileIdsToBeMerged:MergeProfiles', mergeProfiles_profileIdsToBeMerged - The identifier of the profile to be merged into MainProfileId.

Request Lenses

mergeProfiles_fieldSourceProfileIds :: Lens' MergeProfiles (Maybe FieldSourceProfileIds)

The identifiers of the fields in the profile that has the information you want to apply to the merge. For example, say you want to merge EmailAddress from Profile1 into MainProfile. This would be the identifier of the EmailAddress field in Profile1.

mergeProfiles_domainName :: Lens' MergeProfiles Text

The unique name of the domain.

mergeProfiles_mainProfileId :: Lens' MergeProfiles Text

The identifier of the profile to be taken.

mergeProfiles_profileIdsToBeMerged :: Lens' MergeProfiles (NonEmpty Text)

The identifier of the profile to be merged into MainProfileId.

Destructuring the Response

data MergeProfilesResponse

See: newMergeProfilesResponse smart constructor.

Constructors

MergeProfilesResponse'

Fields

message :: Maybe Text

A message that indicates the merge request is complete.

httpStatus :: Int

The response's http status code.

newMergeProfilesResponse

Arguments

:: Int
$sel:httpStatus:MergeProfilesResponse'

-> MergeProfilesResponse

Create a value of MergeProfilesResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:message:MergeProfilesResponse', mergeProfilesResponse_message - A message that indicates the merge request is complete.

$sel:httpStatus:MergeProfilesResponse', mergeProfilesResponse_httpStatus - The response's http status code.

Response Lenses

mergeProfilesResponse_message :: Lens' MergeProfilesResponse (Maybe Text)

A message that indicates the merge request is complete.

mergeProfilesResponse_httpStatus :: Lens' MergeProfilesResponse Int

The response's http status code.