Amazonka.CustomerProfiles.PutProfileObjectType

Description

Defines a ProfileObjectType.

Creating a Request

data PutProfileObjectType

See: newPutProfileObjectType smart constructor.

Constructors

PutProfileObjectType'

Fields

expirationDays :: Maybe Natural

The number of days until the data in the object expires.

templateId :: Maybe Text

A unique identifier for the object template.

keys :: Maybe (HashMap Text [ObjectTypeKey])

A list of unique keys that can be used to map data to the profile.

encryptionKey :: Maybe Text

The customer-provided key to encrypt the profile object that will be created in this profile object type.

allowProfileCreation :: Maybe Bool

Indicates whether a profile should be created when data is received if one doesn’t exist for an object of this type. The default is FALSE. If the AllowProfileCreation flag is set to FALSE, then the service tries to fetch a standard profile and associate this object with the profile. If it is set to TRUE, and if no match is found, then the service creates a new standard profile.

fields :: Maybe (HashMap Text ObjectTypeField)

A map of the name and ObjectType field.

tags :: Maybe (HashMap Text Text)

The tags used to organize, track, or control access for this resource.

domainName :: Text

The unique name of the domain.

objectTypeName :: Text

The name of the profile object type.

description :: Text

Description of the profile object type.

newPutProfileObjectType

Arguments

:: Text
$sel:domainName:PutProfileObjectType'

-> Text
$sel:objectTypeName:PutProfileObjectType'

-> Text
$sel:description:PutProfileObjectType'

-> PutProfileObjectType

Create a value of PutProfileObjectType with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:expirationDays:PutProfileObjectType', putProfileObjectType_expirationDays - The number of days until the data in the object expires.

$sel:templateId:PutProfileObjectType', putProfileObjectType_templateId - A unique identifier for the object template.

$sel:keys:PutProfileObjectType', putProfileObjectType_keys - A list of unique keys that can be used to map data to the profile.

$sel:encryptionKey:PutProfileObjectType', putProfileObjectType_encryptionKey - The customer-provided key to encrypt the profile object that will be created in this profile object type.

$sel:allowProfileCreation:PutProfileObjectType', putProfileObjectType_allowProfileCreation - Indicates whether a profile should be created when data is received if one doesn’t exist for an object of this type. The default is FALSE. If the AllowProfileCreation flag is set to FALSE, then the service tries to fetch a standard profile and associate this object with the profile. If it is set to TRUE, and if no match is found, then the service creates a new standard profile.

$sel:fields:PutProfileObjectType', putProfileObjectType_fields - A map of the name and ObjectType field.

$sel:tags:PutProfileObjectType', putProfileObjectType_tags - The tags used to organize, track, or control access for this resource.

$sel:domainName:PutProfileObjectType', putProfileObjectType_domainName - The unique name of the domain.

$sel:objectTypeName:PutProfileObjectType', putProfileObjectType_objectTypeName - The name of the profile object type.

$sel:description:PutProfileObjectType', putProfileObjectType_description - Description of the profile object type.

Request Lenses

putProfileObjectType_expirationDays :: Lens' PutProfileObjectType (Maybe Natural)

The number of days until the data in the object expires.

putProfileObjectType_templateId :: Lens' PutProfileObjectType (Maybe Text)

A unique identifier for the object template.

putProfileObjectType_keys :: Lens' PutProfileObjectType (Maybe (HashMap Text [ObjectTypeKey]))

A list of unique keys that can be used to map data to the profile.

putProfileObjectType_encryptionKey :: Lens' PutProfileObjectType (Maybe Text)

The customer-provided key to encrypt the profile object that will be created in this profile object type.

putProfileObjectType_allowProfileCreation :: Lens' PutProfileObjectType (Maybe Bool)

Indicates whether a profile should be created when data is received if one doesn’t exist for an object of this type. The default is FALSE. If the AllowProfileCreation flag is set to FALSE, then the service tries to fetch a standard profile and associate this object with the profile. If it is set to TRUE, and if no match is found, then the service creates a new standard profile.

putProfileObjectType_fields :: Lens' PutProfileObjectType (Maybe (HashMap Text ObjectTypeField))

A map of the name and ObjectType field.

putProfileObjectType_tags :: Lens' PutProfileObjectType (Maybe (HashMap Text Text))

The tags used to organize, track, or control access for this resource.

putProfileObjectType_domainName :: Lens' PutProfileObjectType Text

The unique name of the domain.

putProfileObjectType_objectTypeName :: Lens' PutProfileObjectType Text

The name of the profile object type.

putProfileObjectType_description :: Lens' PutProfileObjectType Text

Description of the profile object type.

Destructuring the Response

data PutProfileObjectTypeResponse

See: newPutProfileObjectTypeResponse smart constructor.

Constructors

PutProfileObjectTypeResponse'

Fields

expirationDays :: Maybe Natural

The number of days until the data in the object expires.

lastUpdatedAt :: Maybe POSIX

The timestamp of when the domain was most recently edited.

createdAt :: Maybe POSIX

The timestamp of when the domain was created.

templateId :: Maybe Text

A unique identifier for the object template.

keys :: Maybe (HashMap Text [ObjectTypeKey])

A list of unique keys that can be used to map data to the profile.

encryptionKey :: Maybe Text

The customer-provided key to encrypt the profile object that will be created in this profile object type.

allowProfileCreation :: Maybe Bool

Indicates whether a profile should be created when data is received if one doesn’t exist for an object of this type. The default is FALSE. If the AllowProfileCreation flag is set to FALSE, then the service tries to fetch a standard profile and associate this object with the profile. If it is set to TRUE, and if no match is found, then the service creates a new standard profile.

fields :: Maybe (HashMap Text ObjectTypeField)

A map of the name and ObjectType field.

tags :: Maybe (HashMap Text Text)

The tags used to organize, track, or control access for this resource.

httpStatus :: Int

The response's http status code.

objectTypeName :: Text

The name of the profile object type.

description :: Text

Description of the profile object type.

newPutProfileObjectTypeResponse

Arguments

:: Int
$sel:httpStatus:PutProfileObjectTypeResponse'

-> Text
$sel:objectTypeName:PutProfileObjectType'

-> Text
$sel:description:PutProfileObjectType'

-> PutProfileObjectTypeResponse

Create a value of PutProfileObjectTypeResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:expirationDays:PutProfileObjectType', putProfileObjectTypeResponse_expirationDays - The number of days until the data in the object expires.

$sel:lastUpdatedAt:PutProfileObjectTypeResponse', putProfileObjectTypeResponse_lastUpdatedAt - The timestamp of when the domain was most recently edited.

$sel:createdAt:PutProfileObjectTypeResponse', putProfileObjectTypeResponse_createdAt - The timestamp of when the domain was created.

$sel:templateId:PutProfileObjectType', putProfileObjectTypeResponse_templateId - A unique identifier for the object template.

$sel:keys:PutProfileObjectType', putProfileObjectTypeResponse_keys - A list of unique keys that can be used to map data to the profile.

$sel:encryptionKey:PutProfileObjectType', putProfileObjectTypeResponse_encryptionKey - The customer-provided key to encrypt the profile object that will be created in this profile object type.

$sel:allowProfileCreation:PutProfileObjectType', putProfileObjectTypeResponse_allowProfileCreation - Indicates whether a profile should be created when data is received if one doesn’t exist for an object of this type. The default is FALSE. If the AllowProfileCreation flag is set to FALSE, then the service tries to fetch a standard profile and associate this object with the profile. If it is set to TRUE, and if no match is found, then the service creates a new standard profile.

$sel:fields:PutProfileObjectType', putProfileObjectTypeResponse_fields - A map of the name and ObjectType field.

$sel:tags:PutProfileObjectType', putProfileObjectTypeResponse_tags - The tags used to organize, track, or control access for this resource.

$sel:httpStatus:PutProfileObjectTypeResponse', putProfileObjectTypeResponse_httpStatus - The response's http status code.

$sel:objectTypeName:PutProfileObjectType', putProfileObjectTypeResponse_objectTypeName - The name of the profile object type.

$sel:description:PutProfileObjectType', putProfileObjectTypeResponse_description - Description of the profile object type.

Response Lenses

putProfileObjectTypeResponse_expirationDays :: Lens' PutProfileObjectTypeResponse (Maybe Natural)

The number of days until the data in the object expires.

putProfileObjectTypeResponse_lastUpdatedAt :: Lens' PutProfileObjectTypeResponse (Maybe UTCTime)

The timestamp of when the domain was most recently edited.

putProfileObjectTypeResponse_createdAt :: Lens' PutProfileObjectTypeResponse (Maybe UTCTime)

The timestamp of when the domain was created.

putProfileObjectTypeResponse_templateId :: Lens' PutProfileObjectTypeResponse (Maybe Text)

A unique identifier for the object template.

putProfileObjectTypeResponse_keys :: Lens' PutProfileObjectTypeResponse (Maybe (HashMap Text [ObjectTypeKey]))

A list of unique keys that can be used to map data to the profile.

putProfileObjectTypeResponse_encryptionKey :: Lens' PutProfileObjectTypeResponse (Maybe Text)

The customer-provided key to encrypt the profile object that will be created in this profile object type.

putProfileObjectTypeResponse_allowProfileCreation :: Lens' PutProfileObjectTypeResponse (Maybe Bool)

Indicates whether a profile should be created when data is received if one doesn’t exist for an object of this type. The default is FALSE. If the AllowProfileCreation flag is set to FALSE, then the service tries to fetch a standard profile and associate this object with the profile. If it is set to TRUE, and if no match is found, then the service creates a new standard profile.

putProfileObjectTypeResponse_fields :: Lens' PutProfileObjectTypeResponse (Maybe (HashMap Text ObjectTypeField))

A map of the name and ObjectType field.

putProfileObjectTypeResponse_tags :: Lens' PutProfileObjectTypeResponse (Maybe (HashMap Text Text))

The tags used to organize, track, or control access for this resource.

putProfileObjectTypeResponse_httpStatus :: Lens' PutProfileObjectTypeResponse Int

The response's http status code.

putProfileObjectTypeResponse_objectTypeName :: Lens' PutProfileObjectTypeResponse Text

The name of the profile object type.

putProfileObjectTypeResponse_description :: Lens' PutProfileObjectTypeResponse Text

Description of the profile object type.