Amazonka.CustomerProfiles.SearchProfiles

Description

Searches for profiles within a specific domain name using name, phone number, email address, account number, or a custom defined index.

Creating a Request

data SearchProfiles

See: newSearchProfiles smart constructor.

Constructors

SearchProfiles'

Fields

nextToken :: Maybe Text

The pagination token from the previous SearchProfiles API call.

maxResults :: Maybe Natural

The maximum number of objects returned per page.

domainName :: Text

The unique name of the domain.

keyName :: Text

A searchable identifier of a customer profile. The predefined keys you can use to search include: _account, _profileId, _fullName, _phone, _email, _ctrContactId, _marketoLeadId, _salesforceAccountId, _salesforceContactId, _zendeskUserId, _zendeskExternalId, _serviceNowSystemId.

values :: [Text]

A list of key values.

newSearchProfiles

Arguments

:: Text
$sel:domainName:SearchProfiles'

-> Text
$sel:keyName:SearchProfiles'

-> SearchProfiles

Create a value of SearchProfiles with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:nextToken:SearchProfiles', searchProfiles_nextToken - The pagination token from the previous SearchProfiles API call.

$sel:maxResults:SearchProfiles', searchProfiles_maxResults - The maximum number of objects returned per page.

$sel:domainName:SearchProfiles', searchProfiles_domainName - The unique name of the domain.

$sel:keyName:SearchProfiles', searchProfiles_keyName - A searchable identifier of a customer profile. The predefined keys you can use to search include: _account, _profileId, _fullName, _phone, _email, _ctrContactId, _marketoLeadId, _salesforceAccountId, _salesforceContactId, _zendeskUserId, _zendeskExternalId, _serviceNowSystemId.

$sel:values:SearchProfiles', searchProfiles_values - A list of key values.

Request Lenses

searchProfiles_nextToken :: Lens' SearchProfiles (Maybe Text)

The pagination token from the previous SearchProfiles API call.

searchProfiles_maxResults :: Lens' SearchProfiles (Maybe Natural)

The maximum number of objects returned per page.

searchProfiles_domainName :: Lens' SearchProfiles Text

The unique name of the domain.

searchProfiles_keyName :: Lens' SearchProfiles Text

A searchable identifier of a customer profile. The predefined keys you can use to search include: _account, _profileId, _fullName, _phone, _email, _ctrContactId, _marketoLeadId, _salesforceAccountId, _salesforceContactId, _zendeskUserId, _zendeskExternalId, _serviceNowSystemId.

searchProfiles_values :: Lens' SearchProfiles [Text]

A list of key values.

Destructuring the Response

data SearchProfilesResponse

See: newSearchProfilesResponse smart constructor.

Constructors

SearchProfilesResponse'

Fields

items :: Maybe [Profile]

The list of SearchProfiles instances.

nextToken :: Maybe Text

The pagination token from the previous SearchProfiles API call.

httpStatus :: Int

The response's http status code.

newSearchProfilesResponse

Arguments

:: Int
$sel:httpStatus:SearchProfilesResponse'

-> SearchProfilesResponse

Create a value of SearchProfilesResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:items:SearchProfilesResponse', searchProfilesResponse_items - The list of SearchProfiles instances.

$sel:nextToken:SearchProfiles', searchProfilesResponse_nextToken - The pagination token from the previous SearchProfiles API call.

$sel:httpStatus:SearchProfilesResponse', searchProfilesResponse_httpStatus - The response's http status code.

Response Lenses

searchProfilesResponse_items :: Lens' SearchProfilesResponse (Maybe [Profile])

The list of SearchProfiles instances.

searchProfilesResponse_nextToken :: Lens' SearchProfilesResponse (Maybe Text)

The pagination token from the previous SearchProfiles API call.

searchProfilesResponse_httpStatus :: Lens' SearchProfilesResponse Int

The response's http status code.