Amazonka.CustomerProfiles.Types.Address

Description

Documentation

data Address

A generic address associated with the customer that is not mailing, shipping, or billing.

See: newAddress smart constructor.

Constructors

Address'

Fields

address2 :: Maybe Text

The second line of a customer address.

state :: Maybe Text

The state in which a customer lives.

address4 :: Maybe Text

The fourth line of a customer address.

county :: Maybe Text

The county in which a customer lives.

address3 :: Maybe Text

The third line of a customer address.

postalCode :: Maybe Text

The postal code of a customer address.

country :: Maybe Text

The country in which a customer lives.

city :: Maybe Text

The city in which a customer lives.

address1 :: Maybe Text

The first line of a customer address.

province :: Maybe Text

The province in which a customer lives.

newAddress :: Address

Create a value of Address with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:address2:Address', address_address2 - The second line of a customer address.

$sel:state:Address', address_state - The state in which a customer lives.

$sel:address4:Address', address_address4 - The fourth line of a customer address.

$sel:county:Address', address_county - The county in which a customer lives.

$sel:address3:Address', address_address3 - The third line of a customer address.

$sel:postalCode:Address', address_postalCode - The postal code of a customer address.

$sel:country:Address', address_country - The country in which a customer lives.

$sel:city:Address', address_city - The city in which a customer lives.

$sel:address1:Address', address_address1 - The first line of a customer address.

$sel:province:Address', address_province - The province in which a customer lives.

address_address2 :: Lens' Address (Maybe Text)

The second line of a customer address.

address_state :: Lens' Address (Maybe Text)

The state in which a customer lives.

address_address4 :: Lens' Address (Maybe Text)

The fourth line of a customer address.

address_county :: Lens' Address (Maybe Text)

The county in which a customer lives.

address_address3 :: Lens' Address (Maybe Text)

The third line of a customer address.

address_postalCode :: Lens' Address (Maybe Text)

The postal code of a customer address.

address_country :: Lens' Address (Maybe Text)

The country in which a customer lives.

address_city :: Lens' Address (Maybe Text)

The city in which a customer lives.

address_address1 :: Lens' Address (Maybe Text)

The first line of a customer address.

address_province :: Lens' Address (Maybe Text)

The province in which a customer lives.