Amazonka.CustomerProfiles.Types.FieldSourceProfileIds

Description

Documentation

data FieldSourceProfileIds

A duplicate customer profile that is to be merged into a main profile.

See: newFieldSourceProfileIds smart constructor.

Constructors

FieldSourceProfileIds'

Fields

shippingAddress :: Maybe Text

A unique identifier for the shipping address field to be merged.

mobilePhoneNumber :: Maybe Text

A unique identifier for the mobile phone number field to be merged.

mailingAddress :: Maybe Text

A unique identifier for the mailing address field to be merged.

middleName :: Maybe Text

A unique identifier for the middle name field to be merged.

personalEmailAddress :: Maybe Text

A unique identifier for the personal email address field to be merged.

lastName :: Maybe Text

A unique identifier for the last name field to be merged.

additionalInformation :: Maybe Text

A unique identifier for the additional information field to be merged.

homePhoneNumber :: Maybe Text

A unique identifier for the home phone number field to be merged.

address :: Maybe Text

A unique identifier for the party type field to be merged.

partyType :: Maybe Text

A unique identifier for the party type field to be merged.

businessEmailAddress :: Maybe Text

A unique identifier for the party type field to be merged.

attributes :: Maybe (HashMap Text Text)

A unique identifier for the attributes field to be merged.

gender :: Maybe Text

A unique identifier for the gender field to be merged.

phoneNumber :: Maybe Text

A unique identifier for the phone number field to be merged.

accountNumber :: Maybe Text

A unique identifier for the account number field to be merged.

emailAddress :: Maybe Text

A unique identifier for the email address field to be merged.

firstName :: Maybe Text

A unique identifier for the first name field to be merged.

billingAddress :: Maybe Text

A unique identifier for the billing type field to be merged.

businessPhoneNumber :: Maybe Text

A unique identifier for the business phone number field to be merged.

birthDate :: Maybe Text

A unique identifier for the birthdate field to be merged.

businessName :: Maybe Text

A unique identifier for the business name field to be merged.

newFieldSourceProfileIds :: FieldSourceProfileIds

Create a value of FieldSourceProfileIds with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:shippingAddress:FieldSourceProfileIds', fieldSourceProfileIds_shippingAddress - A unique identifier for the shipping address field to be merged.

$sel:mobilePhoneNumber:FieldSourceProfileIds', fieldSourceProfileIds_mobilePhoneNumber - A unique identifier for the mobile phone number field to be merged.

$sel:mailingAddress:FieldSourceProfileIds', fieldSourceProfileIds_mailingAddress - A unique identifier for the mailing address field to be merged.

$sel:middleName:FieldSourceProfileIds', fieldSourceProfileIds_middleName - A unique identifier for the middle name field to be merged.

$sel:personalEmailAddress:FieldSourceProfileIds', fieldSourceProfileIds_personalEmailAddress - A unique identifier for the personal email address field to be merged.

$sel:lastName:FieldSourceProfileIds', fieldSourceProfileIds_lastName - A unique identifier for the last name field to be merged.

$sel:additionalInformation:FieldSourceProfileIds', fieldSourceProfileIds_additionalInformation - A unique identifier for the additional information field to be merged.

$sel:homePhoneNumber:FieldSourceProfileIds', fieldSourceProfileIds_homePhoneNumber - A unique identifier for the home phone number field to be merged.

$sel:address:FieldSourceProfileIds', fieldSourceProfileIds_address - A unique identifier for the party type field to be merged.

$sel:partyType:FieldSourceProfileIds', fieldSourceProfileIds_partyType - A unique identifier for the party type field to be merged.

$sel:businessEmailAddress:FieldSourceProfileIds', fieldSourceProfileIds_businessEmailAddress - A unique identifier for the party type field to be merged.

$sel:attributes:FieldSourceProfileIds', fieldSourceProfileIds_attributes - A unique identifier for the attributes field to be merged.

$sel:gender:FieldSourceProfileIds', fieldSourceProfileIds_gender - A unique identifier for the gender field to be merged.

$sel:phoneNumber:FieldSourceProfileIds', fieldSourceProfileIds_phoneNumber - A unique identifier for the phone number field to be merged.

$sel:accountNumber:FieldSourceProfileIds', fieldSourceProfileIds_accountNumber - A unique identifier for the account number field to be merged.

$sel:emailAddress:FieldSourceProfileIds', fieldSourceProfileIds_emailAddress - A unique identifier for the email address field to be merged.

$sel:firstName:FieldSourceProfileIds', fieldSourceProfileIds_firstName - A unique identifier for the first name field to be merged.

$sel:billingAddress:FieldSourceProfileIds', fieldSourceProfileIds_billingAddress - A unique identifier for the billing type field to be merged.

$sel:businessPhoneNumber:FieldSourceProfileIds', fieldSourceProfileIds_businessPhoneNumber - A unique identifier for the business phone number field to be merged.

$sel:birthDate:FieldSourceProfileIds', fieldSourceProfileIds_birthDate - A unique identifier for the birthdate field to be merged.

$sel:businessName:FieldSourceProfileIds', fieldSourceProfileIds_businessName - A unique identifier for the business name field to be merged.

fieldSourceProfileIds_shippingAddress :: Lens' FieldSourceProfileIds (Maybe Text)

A unique identifier for the shipping address field to be merged.

fieldSourceProfileIds_mobilePhoneNumber :: Lens' FieldSourceProfileIds (Maybe Text)

A unique identifier for the mobile phone number field to be merged.

fieldSourceProfileIds_mailingAddress :: Lens' FieldSourceProfileIds (Maybe Text)

A unique identifier for the mailing address field to be merged.

fieldSourceProfileIds_middleName :: Lens' FieldSourceProfileIds (Maybe Text)

A unique identifier for the middle name field to be merged.

fieldSourceProfileIds_personalEmailAddress :: Lens' FieldSourceProfileIds (Maybe Text)

A unique identifier for the personal email address field to be merged.

fieldSourceProfileIds_lastName :: Lens' FieldSourceProfileIds (Maybe Text)

A unique identifier for the last name field to be merged.

fieldSourceProfileIds_additionalInformation :: Lens' FieldSourceProfileIds (Maybe Text)

A unique identifier for the additional information field to be merged.

fieldSourceProfileIds_homePhoneNumber :: Lens' FieldSourceProfileIds (Maybe Text)

A unique identifier for the home phone number field to be merged.

fieldSourceProfileIds_address :: Lens' FieldSourceProfileIds (Maybe Text)

A unique identifier for the party type field to be merged.

fieldSourceProfileIds_partyType :: Lens' FieldSourceProfileIds (Maybe Text)

A unique identifier for the party type field to be merged.

fieldSourceProfileIds_businessEmailAddress :: Lens' FieldSourceProfileIds (Maybe Text)

A unique identifier for the party type field to be merged.

fieldSourceProfileIds_attributes :: Lens' FieldSourceProfileIds (Maybe (HashMap Text Text))

A unique identifier for the attributes field to be merged.

fieldSourceProfileIds_gender :: Lens' FieldSourceProfileIds (Maybe Text)

A unique identifier for the gender field to be merged.

fieldSourceProfileIds_phoneNumber :: Lens' FieldSourceProfileIds (Maybe Text)

A unique identifier for the phone number field to be merged.

fieldSourceProfileIds_accountNumber :: Lens' FieldSourceProfileIds (Maybe Text)

A unique identifier for the account number field to be merged.

fieldSourceProfileIds_emailAddress :: Lens' FieldSourceProfileIds (Maybe Text)

A unique identifier for the email address field to be merged.

fieldSourceProfileIds_firstName :: Lens' FieldSourceProfileIds (Maybe Text)

A unique identifier for the first name field to be merged.

fieldSourceProfileIds_billingAddress :: Lens' FieldSourceProfileIds (Maybe Text)

A unique identifier for the billing type field to be merged.

fieldSourceProfileIds_businessPhoneNumber :: Lens' FieldSourceProfileIds (Maybe Text)

A unique identifier for the business phone number field to be merged.

fieldSourceProfileIds_birthDate :: Lens' FieldSourceProfileIds (Maybe Text)

A unique identifier for the birthdate field to be merged.

fieldSourceProfileIds_businessName :: Lens' FieldSourceProfileIds (Maybe Text)

A unique identifier for the business name field to be merged.