Amazonka.CustomerProfiles.Types.Profile

Description

Documentation

data Profile

The standard profile of a customer.

See: newProfile smart constructor.

Constructors

Profile'

Fields

shippingAddress :: Maybe Address

The customer’s shipping address.

mobilePhoneNumber :: Maybe Text

The customer’s mobile phone number.

mailingAddress :: Maybe Address

The customer’s mailing address.

middleName :: Maybe Text

The customer’s middle name.

personalEmailAddress :: Maybe Text

The customer’s personal email address.

lastName :: Maybe Text

The customer’s last name.

additionalInformation :: Maybe Text

Any additional information relevant to the customer’s profile.

homePhoneNumber :: Maybe Text

The customer’s home phone number.

address :: Maybe Address

A generic address associated with the customer that is not mailing, shipping, or billing.

partyType :: Maybe PartyType

The type of profile used to describe the customer.

profileId :: Maybe Text

The unique identifier of a customer profile.

businessEmailAddress :: Maybe Text

The customer’s business email address.

attributes :: Maybe (HashMap Text Text)

A key value pair of attributes of a customer profile.

gender :: Maybe Gender

The gender with which the customer identifies.

phoneNumber :: Maybe Text

The customer's phone number, which has not been specified as a mobile, home, or business number.

accountNumber :: Maybe Text

A unique account number that you have given to the customer.

emailAddress :: Maybe Text

The customer’s email address, which has not been specified as a personal or business address.

firstName :: Maybe Text

The customer’s first name.

billingAddress :: Maybe Address

The customer’s billing address.

businessPhoneNumber :: Maybe Text

The customer’s home phone number.

birthDate :: Maybe Text

The customer’s birth date.

businessName :: Maybe Text

The name of the customer’s business.

newProfile :: Profile

Create a value of Profile with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:shippingAddress:Profile', profile_shippingAddress - The customer’s shipping address.

$sel:mobilePhoneNumber:Profile', profile_mobilePhoneNumber - The customer’s mobile phone number.

$sel:mailingAddress:Profile', profile_mailingAddress - The customer’s mailing address.

$sel:middleName:Profile', profile_middleName - The customer’s middle name.

$sel:personalEmailAddress:Profile', profile_personalEmailAddress - The customer’s personal email address.

$sel:lastName:Profile', profile_lastName - The customer’s last name.

$sel:additionalInformation:Profile', profile_additionalInformation - Any additional information relevant to the customer’s profile.

$sel:homePhoneNumber:Profile', profile_homePhoneNumber - The customer’s home phone number.

$sel:address:Profile', profile_address - A generic address associated with the customer that is not mailing, shipping, or billing.

$sel:partyType:Profile', profile_partyType - The type of profile used to describe the customer.

$sel:profileId:Profile', profile_profileId - The unique identifier of a customer profile.

$sel:businessEmailAddress:Profile', profile_businessEmailAddress - The customer’s business email address.

$sel:attributes:Profile', profile_attributes - A key value pair of attributes of a customer profile.

$sel:gender:Profile', profile_gender - The gender with which the customer identifies.

$sel:phoneNumber:Profile', profile_phoneNumber - The customer's phone number, which has not been specified as a mobile, home, or business number.

$sel:accountNumber:Profile', profile_accountNumber - A unique account number that you have given to the customer.

$sel:emailAddress:Profile', profile_emailAddress - The customer’s email address, which has not been specified as a personal or business address.

$sel:firstName:Profile', profile_firstName - The customer’s first name.

$sel:billingAddress:Profile', profile_billingAddress - The customer’s billing address.

$sel:businessPhoneNumber:Profile', profile_businessPhoneNumber - The customer’s home phone number.

$sel:birthDate:Profile', profile_birthDate - The customer’s birth date.

$sel:businessName:Profile', profile_businessName - The name of the customer’s business.

profile_shippingAddress :: Lens' Profile (Maybe Address)

The customer’s shipping address.

profile_mobilePhoneNumber :: Lens' Profile (Maybe Text)

The customer’s mobile phone number.

profile_mailingAddress :: Lens' Profile (Maybe Address)

The customer’s mailing address.

profile_middleName :: Lens' Profile (Maybe Text)

The customer’s middle name.

profile_personalEmailAddress :: Lens' Profile (Maybe Text)

The customer’s personal email address.

profile_lastName :: Lens' Profile (Maybe Text)

The customer’s last name.

profile_additionalInformation :: Lens' Profile (Maybe Text)

Any additional information relevant to the customer’s profile.

profile_homePhoneNumber :: Lens' Profile (Maybe Text)

The customer’s home phone number.

profile_address :: Lens' Profile (Maybe Address)

A generic address associated with the customer that is not mailing, shipping, or billing.

profile_partyType :: Lens' Profile (Maybe PartyType)

The type of profile used to describe the customer.

profile_profileId :: Lens' Profile (Maybe Text)

The unique identifier of a customer profile.

profile_businessEmailAddress :: Lens' Profile (Maybe Text)

The customer’s business email address.

profile_attributes :: Lens' Profile (Maybe (HashMap Text Text))

A key value pair of attributes of a customer profile.

profile_gender :: Lens' Profile (Maybe Gender)

The gender with which the customer identifies.

profile_phoneNumber :: Lens' Profile (Maybe Text)

The customer's phone number, which has not been specified as a mobile, home, or business number.

profile_accountNumber :: Lens' Profile (Maybe Text)

A unique account number that you have given to the customer.

profile_emailAddress :: Lens' Profile (Maybe Text)

The customer’s email address, which has not been specified as a personal or business address.

profile_firstName :: Lens' Profile (Maybe Text)

The customer’s first name.

profile_billingAddress :: Lens' Profile (Maybe Address)

The customer’s billing address.

profile_businessPhoneNumber :: Lens' Profile (Maybe Text)

The customer’s home phone number.

profile_birthDate :: Lens' Profile (Maybe Text)

The customer’s birth date.

profile_businessName :: Lens' Profile (Maybe Text)

The name of the customer’s business.