Amazonka.CustomerProfiles.Types.UpdateAddress

Description

Documentation

data UpdateAddress

Updates associated with the address properties of a customer profile.

See: newUpdateAddress smart constructor.

Constructors

UpdateAddress'

Fields

address2 :: Maybe Text

The second line of a customer address.

state :: Maybe Text

The state in which a customer lives.

address4 :: Maybe Text

The fourth line of a customer address.

county :: Maybe Text

The county in which a customer lives.

address3 :: Maybe Text

The third line of a customer address.

postalCode :: Maybe Text

The postal code of a customer address.

country :: Maybe Text

The country in which a customer lives.

city :: Maybe Text

The city in which a customer lives.

address1 :: Maybe Text

The first line of a customer address.

province :: Maybe Text

The province in which a customer lives.

newUpdateAddress :: UpdateAddress

Create a value of UpdateAddress with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:address2:UpdateAddress', updateAddress_address2 - The second line of a customer address.

$sel:state:UpdateAddress', updateAddress_state - The state in which a customer lives.

$sel:address4:UpdateAddress', updateAddress_address4 - The fourth line of a customer address.

$sel:county:UpdateAddress', updateAddress_county - The county in which a customer lives.

$sel:address3:UpdateAddress', updateAddress_address3 - The third line of a customer address.

$sel:postalCode:UpdateAddress', updateAddress_postalCode - The postal code of a customer address.

$sel:country:UpdateAddress', updateAddress_country - The country in which a customer lives.

$sel:city:UpdateAddress', updateAddress_city - The city in which a customer lives.

$sel:address1:UpdateAddress', updateAddress_address1 - The first line of a customer address.

$sel:province:UpdateAddress', updateAddress_province - The province in which a customer lives.

updateAddress_address2 :: Lens' UpdateAddress (Maybe Text)

The second line of a customer address.

updateAddress_state :: Lens' UpdateAddress (Maybe Text)

The state in which a customer lives.

updateAddress_address4 :: Lens' UpdateAddress (Maybe Text)

The fourth line of a customer address.

updateAddress_county :: Lens' UpdateAddress (Maybe Text)

The county in which a customer lives.

updateAddress_address3 :: Lens' UpdateAddress (Maybe Text)

The third line of a customer address.

updateAddress_postalCode :: Lens' UpdateAddress (Maybe Text)

The postal code of a customer address.

updateAddress_country :: Lens' UpdateAddress (Maybe Text)

The country in which a customer lives.

updateAddress_city :: Lens' UpdateAddress (Maybe Text)

The city in which a customer lives.

updateAddress_address1 :: Lens' UpdateAddress (Maybe Text)

The first line of a customer address.

updateAddress_province :: Lens' UpdateAddress (Maybe Text)

The province in which a customer lives.