Amazonka.CustomerProfiles.UpdateDomain

Description

Updates the properties of a domain, including creating or selecting a dead letter queue or an encryption key.

After a domain is created, the name can’t be changed.

Use this API or CreateDomain to enable identity resolution: set Matching to true.

Creating a Request

data UpdateDomain

See: newUpdateDomain smart constructor.

Constructors

UpdateDomain'

Fields

defaultExpirationDays :: Maybe Natural

The default number of days until the data within the domain expires.

defaultEncryptionKey :: Maybe Text

The default encryption key, which is an AWS managed key, is used when no specific type of encryption key is specified. It is used to encrypt all data before it is placed in permanent or semi-permanent storage. If specified as an empty string, it will clear any existing value.

matching :: Maybe MatchingRequest

The process of matching duplicate profiles. If Matching = true, Amazon Connect Customer Profiles starts a weekly batch process every Saturday at 12AM UTC to detect duplicate profiles in your domains. After that batch process completes, use the GetMatches API to return and review the results.

deadLetterQueueUrl :: Maybe Text

The URL of the SQS dead letter queue, which is used for reporting errors associated with ingesting data from third party applications. If specified as an empty string, it will clear any existing value. You must set up a policy on the DeadLetterQueue for the SendMessage operation to enable Amazon Connect Customer Profiles to send messages to the DeadLetterQueue.

tags :: Maybe (HashMap Text Text)

The tags used to organize, track, or control access for this resource.

domainName :: Text

The unique name of the domain.

newUpdateDomain

Arguments

:: Text
$sel:domainName:UpdateDomain'

-> UpdateDomain

Create a value of UpdateDomain with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:defaultExpirationDays:UpdateDomain', updateDomain_defaultExpirationDays - The default number of days until the data within the domain expires.

$sel:defaultEncryptionKey:UpdateDomain', updateDomain_defaultEncryptionKey - The default encryption key, which is an AWS managed key, is used when no specific type of encryption key is specified. It is used to encrypt all data before it is placed in permanent or semi-permanent storage. If specified as an empty string, it will clear any existing value.

$sel:matching:UpdateDomain', updateDomain_matching - The process of matching duplicate profiles. If Matching = true, Amazon Connect Customer Profiles starts a weekly batch process every Saturday at 12AM UTC to detect duplicate profiles in your domains. After that batch process completes, use the GetMatches API to return and review the results.

$sel:deadLetterQueueUrl:UpdateDomain', updateDomain_deadLetterQueueUrl - The URL of the SQS dead letter queue, which is used for reporting errors associated with ingesting data from third party applications. If specified as an empty string, it will clear any existing value. You must set up a policy on the DeadLetterQueue for the SendMessage operation to enable Amazon Connect Customer Profiles to send messages to the DeadLetterQueue.

$sel:tags:UpdateDomain', updateDomain_tags - The tags used to organize, track, or control access for this resource.

$sel:domainName:UpdateDomain', updateDomain_domainName - The unique name of the domain.

Request Lenses

updateDomain_defaultExpirationDays :: Lens' UpdateDomain (Maybe Natural)

The default number of days until the data within the domain expires.

updateDomain_defaultEncryptionKey :: Lens' UpdateDomain (Maybe Text)

The default encryption key, which is an AWS managed key, is used when no specific type of encryption key is specified. It is used to encrypt all data before it is placed in permanent or semi-permanent storage. If specified as an empty string, it will clear any existing value.

updateDomain_matching :: Lens' UpdateDomain (Maybe MatchingRequest)

The process of matching duplicate profiles. If Matching = true, Amazon Connect Customer Profiles starts a weekly batch process every Saturday at 12AM UTC to detect duplicate profiles in your domains. After that batch process completes, use the GetMatches API to return and review the results.

updateDomain_deadLetterQueueUrl :: Lens' UpdateDomain (Maybe Text)

The URL of the SQS dead letter queue, which is used for reporting errors associated with ingesting data from third party applications. If specified as an empty string, it will clear any existing value. You must set up a policy on the DeadLetterQueue for the SendMessage operation to enable Amazon Connect Customer Profiles to send messages to the DeadLetterQueue.

updateDomain_tags :: Lens' UpdateDomain (Maybe (HashMap Text Text))

The tags used to organize, track, or control access for this resource.

updateDomain_domainName :: Lens' UpdateDomain Text

The unique name of the domain.

Destructuring the Response

data UpdateDomainResponse

See: newUpdateDomainResponse smart constructor.

Constructors

UpdateDomainResponse'

Fields

defaultExpirationDays :: Maybe Natural

The default number of days until the data within the domain expires.

defaultEncryptionKey :: Maybe Text

The default encryption key, which is an AWS managed key, is used when no specific type of encryption key is specified. It is used to encrypt all data before it is placed in permanent or semi-permanent storage.

matching :: Maybe MatchingResponse

The process of matching duplicate profiles. If Matching = true, Amazon Connect Customer Profiles starts a weekly batch process every Saturday at 12AM UTC to detect duplicate profiles in your domains. After that batch process completes, use the GetMatches API to return and review the results.

deadLetterQueueUrl :: Maybe Text

The URL of the SQS dead letter queue, which is used for reporting errors associated with ingesting data from third party applications.

tags :: Maybe (HashMap Text Text)

The tags used to organize, track, or control access for this resource.

httpStatus :: Int

The response's http status code.

domainName :: Text

The unique name of the domain.

createdAt :: POSIX

The timestamp of when the domain was created.

lastUpdatedAt :: POSIX

The timestamp of when the domain was most recently edited.

newUpdateDomainResponse

Arguments

:: Int
$sel:httpStatus:UpdateDomainResponse'

-> Text
$sel:domainName:UpdateDomain'

-> UTCTime
$sel:createdAt:UpdateDomainResponse'

-> UTCTime
$sel:lastUpdatedAt:UpdateDomainResponse'

-> UpdateDomainResponse

Create a value of UpdateDomainResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:defaultExpirationDays:UpdateDomain', updateDomainResponse_defaultExpirationDays - The default number of days until the data within the domain expires.

$sel:defaultEncryptionKey:UpdateDomain', updateDomainResponse_defaultEncryptionKey - The default encryption key, which is an AWS managed key, is used when no specific type of encryption key is specified. It is used to encrypt all data before it is placed in permanent or semi-permanent storage.

$sel:matching:UpdateDomain', updateDomainResponse_matching - The process of matching duplicate profiles. If Matching = true, Amazon Connect Customer Profiles starts a weekly batch process every Saturday at 12AM UTC to detect duplicate profiles in your domains. After that batch process completes, use the GetMatches API to return and review the results.

$sel:deadLetterQueueUrl:UpdateDomain', updateDomainResponse_deadLetterQueueUrl - The URL of the SQS dead letter queue, which is used for reporting errors associated with ingesting data from third party applications.

$sel:tags:UpdateDomain', updateDomainResponse_tags - The tags used to organize, track, or control access for this resource.

$sel:httpStatus:UpdateDomainResponse', updateDomainResponse_httpStatus - The response's http status code.

$sel:domainName:UpdateDomain', updateDomainResponse_domainName - The unique name of the domain.

$sel:createdAt:UpdateDomainResponse', updateDomainResponse_createdAt - The timestamp of when the domain was created.

$sel:lastUpdatedAt:UpdateDomainResponse', updateDomainResponse_lastUpdatedAt - The timestamp of when the domain was most recently edited.

Response Lenses

updateDomainResponse_defaultExpirationDays :: Lens' UpdateDomainResponse (Maybe Natural)

The default number of days until the data within the domain expires.

updateDomainResponse_defaultEncryptionKey :: Lens' UpdateDomainResponse (Maybe Text)

The default encryption key, which is an AWS managed key, is used when no specific type of encryption key is specified. It is used to encrypt all data before it is placed in permanent or semi-permanent storage.

updateDomainResponse_matching :: Lens' UpdateDomainResponse (Maybe MatchingResponse)

The process of matching duplicate profiles. If Matching = true, Amazon Connect Customer Profiles starts a weekly batch process every Saturday at 12AM UTC to detect duplicate profiles in your domains. After that batch process completes, use the GetMatches API to return and review the results.

updateDomainResponse_deadLetterQueueUrl :: Lens' UpdateDomainResponse (Maybe Text)

The URL of the SQS dead letter queue, which is used for reporting errors associated with ingesting data from third party applications.

updateDomainResponse_tags :: Lens' UpdateDomainResponse (Maybe (HashMap Text Text))

The tags used to organize, track, or control access for this resource.

updateDomainResponse_httpStatus :: Lens' UpdateDomainResponse Int

The response's http status code.

updateDomainResponse_domainName :: Lens' UpdateDomainResponse Text

The unique name of the domain.

updateDomainResponse_createdAt :: Lens' UpdateDomainResponse UTCTime

The timestamp of when the domain was created.

updateDomainResponse_lastUpdatedAt :: Lens' UpdateDomainResponse UTCTime

The timestamp of when the domain was most recently edited.