Amazonka.DAX.DeleteCluster

Description

Deletes a previously provisioned DAX cluster. DeleteCluster deletes all associated nodes, node endpoints and the DAX cluster itself. When you receive a successful response from this action, DAX immediately begins deleting the cluster; you cannot cancel or revert this action.

Creating a Request

data DeleteCluster

See: newDeleteCluster smart constructor.

Constructors

DeleteCluster'

Fields

clusterName :: Text

The name of the cluster to be deleted.

newDeleteCluster

Arguments

:: Text
$sel:clusterName:DeleteCluster'

-> DeleteCluster

Create a value of DeleteCluster with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:clusterName:DeleteCluster', deleteCluster_clusterName - The name of the cluster to be deleted.

Request Lenses

deleteCluster_clusterName :: Lens' DeleteCluster Text

The name of the cluster to be deleted.

Destructuring the Response

data DeleteClusterResponse

See: newDeleteClusterResponse smart constructor.

Constructors

DeleteClusterResponse'

Fields

cluster :: Maybe Cluster

A description of the DAX cluster that is being deleted.

httpStatus :: Int

The response's http status code.

newDeleteClusterResponse

Arguments

:: Int
$sel:httpStatus:DeleteClusterResponse'

-> DeleteClusterResponse

Create a value of DeleteClusterResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:cluster:DeleteClusterResponse', deleteClusterResponse_cluster - A description of the DAX cluster that is being deleted.

$sel:httpStatus:DeleteClusterResponse', deleteClusterResponse_httpStatus - The response's http status code.

Response Lenses

deleteClusterResponse_cluster :: Lens' DeleteClusterResponse (Maybe Cluster)

A description of the DAX cluster that is being deleted.

deleteClusterResponse_httpStatus :: Lens' DeleteClusterResponse Int

The response's http status code.