Amazonka.DAX.DescribeEvents

Description

Returns events related to DAX clusters and parameter groups. You can obtain events specific to a particular DAX cluster or parameter group by providing the name as a parameter.

By default, only the events occurring within the last 24 hours are returned; however, you can retrieve up to 14 days' worth of events if necessary.

This operation returns paginated results.

Creating a Request

data DescribeEvents

See: newDescribeEvents smart constructor.

Constructors

DescribeEvents'

Fields

sourceName :: Maybe Text

The identifier of the event source for which events will be returned. If not specified, then all sources are included in the response.

startTime :: Maybe POSIX

The beginning of the time interval to retrieve events for, specified in ISO 8601 format.

sourceType :: Maybe SourceType

The event source to retrieve events for. If no value is specified, all events are returned.

nextToken :: Maybe Text

An optional token returned from a prior request. Use this token for pagination of results from this action. If this parameter is specified, the response includes only results beyond the token, up to the value specified by MaxResults.

endTime :: Maybe POSIX

The end of the time interval for which to retrieve events, specified in ISO 8601 format.

duration :: Maybe Int

The number of minutes' worth of events to retrieve.

maxResults :: Maybe Int

The maximum number of results to include in the response. If more results exist than the specified MaxResults value, a token is included in the response so that the remaining results can be retrieved.

The value for MaxResults must be between 20 and 100.

newDescribeEvents :: DescribeEvents

Create a value of DescribeEvents with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:sourceName:DescribeEvents', describeEvents_sourceName - The identifier of the event source for which events will be returned. If not specified, then all sources are included in the response.

$sel:startTime:DescribeEvents', describeEvents_startTime - The beginning of the time interval to retrieve events for, specified in ISO 8601 format.

$sel:sourceType:DescribeEvents', describeEvents_sourceType - The event source to retrieve events for. If no value is specified, all events are returned.

$sel:nextToken:DescribeEvents', describeEvents_nextToken - An optional token returned from a prior request. Use this token for pagination of results from this action. If this parameter is specified, the response includes only results beyond the token, up to the value specified by MaxResults.

$sel:endTime:DescribeEvents', describeEvents_endTime - The end of the time interval for which to retrieve events, specified in ISO 8601 format.

$sel:duration:DescribeEvents', describeEvents_duration - The number of minutes' worth of events to retrieve.

$sel:maxResults:DescribeEvents', describeEvents_maxResults - The maximum number of results to include in the response. If more results exist than the specified MaxResults value, a token is included in the response so that the remaining results can be retrieved.

The value for MaxResults must be between 20 and 100.

Request Lenses

describeEvents_sourceName :: Lens' DescribeEvents (Maybe Text)

The identifier of the event source for which events will be returned. If not specified, then all sources are included in the response.

describeEvents_startTime :: Lens' DescribeEvents (Maybe UTCTime)

The beginning of the time interval to retrieve events for, specified in ISO 8601 format.

describeEvents_sourceType :: Lens' DescribeEvents (Maybe SourceType)

The event source to retrieve events for. If no value is specified, all events are returned.

describeEvents_nextToken :: Lens' DescribeEvents (Maybe Text)

An optional token returned from a prior request. Use this token for pagination of results from this action. If this parameter is specified, the response includes only results beyond the token, up to the value specified by MaxResults.

describeEvents_endTime :: Lens' DescribeEvents (Maybe UTCTime)

The end of the time interval for which to retrieve events, specified in ISO 8601 format.

describeEvents_duration :: Lens' DescribeEvents (Maybe Int)

The number of minutes' worth of events to retrieve.

describeEvents_maxResults :: Lens' DescribeEvents (Maybe Int)

The maximum number of results to include in the response. If more results exist than the specified MaxResults value, a token is included in the response so that the remaining results can be retrieved.

The value for MaxResults must be between 20 and 100.

Destructuring the Response

data DescribeEventsResponse

See: newDescribeEventsResponse smart constructor.

Constructors

DescribeEventsResponse'

Fields

nextToken :: Maybe Text

Provides an identifier to allow retrieval of paginated results.

events :: Maybe [Event]

An array of events. Each element in the array represents one event.

httpStatus :: Int

The response's http status code.

newDescribeEventsResponse

Arguments

:: Int
$sel:httpStatus:DescribeEventsResponse'

-> DescribeEventsResponse

Create a value of DescribeEventsResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:nextToken:DescribeEvents', describeEventsResponse_nextToken - Provides an identifier to allow retrieval of paginated results.

$sel:events:DescribeEventsResponse', describeEventsResponse_events - An array of events. Each element in the array represents one event.

$sel:httpStatus:DescribeEventsResponse', describeEventsResponse_httpStatus - The response's http status code.

Response Lenses

describeEventsResponse_nextToken :: Lens' DescribeEventsResponse (Maybe Text)

Provides an identifier to allow retrieval of paginated results.

describeEventsResponse_events :: Lens' DescribeEventsResponse (Maybe [Event])

An array of events. Each element in the array represents one event.

describeEventsResponse_httpStatus :: Lens' DescribeEventsResponse Int

The response's http status code.