Amazonka.DAX.Types

Description

Service Configuration

defaultService :: Service

API version 2017-04-19 of the Amazon DynamoDB Accelerator (DAX) SDK configuration.

Errors

_SubnetQuotaExceededFault :: AsError a => Getting (First ServiceError) a ServiceError

The request cannot be processed because it would exceed the allowed number of subnets in a subnet group.

_ParameterGroupNotFoundFault :: AsError a => Getting (First ServiceError) a ServiceError

The specified parameter group does not exist.

_InvalidParameterGroupStateFault :: AsError a => Getting (First ServiceError) a ServiceError

One or more parameters in a parameter group are in an invalid state.

_SubnetGroupInUseFault :: AsError a => Getting (First ServiceError) a ServiceError

The specified subnet group is currently in use.

_ParameterGroupAlreadyExistsFault :: AsError a => Getting (First ServiceError) a ServiceError

The specified parameter group already exists.

_InvalidSubnet :: AsError a => Getting (First ServiceError) a ServiceError

An invalid subnet identifier was specified.

_TagQuotaPerResourceExceeded :: AsError a => Getting (First ServiceError) a ServiceError

You have exceeded the maximum number of tags for this DAX cluster.

_ClusterNotFoundFault :: AsError a => Getting (First ServiceError) a ServiceError

The requested cluster ID does not refer to an existing DAX cluster.

_TagNotFoundFault :: AsError a => Getting (First ServiceError) a ServiceError

The tag does not exist.

_NodeQuotaForClusterExceededFault :: AsError a => Getting (First ServiceError) a ServiceError

You have attempted to exceed the maximum number of nodes for a DAX cluster.

_InvalidClusterStateFault :: AsError a => Getting (First ServiceError) a ServiceError

The requested DAX cluster is not in the available state.

_ServiceLinkedRoleNotFoundFault :: AsError a => Getting (First ServiceError) a ServiceError

The specified service linked role (SLR) was not found.

_InsufficientClusterCapacityFault :: AsError a => Getting (First ServiceError) a ServiceError

There are not enough system resources to create the cluster you requested (or to resize an already-existing cluster).

_NodeNotFoundFault :: AsError a => Getting (First ServiceError) a ServiceError

None of the nodes in the cluster have the given node ID.

_ParameterGroupQuotaExceededFault :: AsError a => Getting (First ServiceError) a ServiceError

You have attempted to exceed the maximum number of parameter groups.

_InvalidParameterValueException :: AsError a => Getting (First ServiceError) a ServiceError

The value for a parameter is invalid.

_InvalidVPCNetworkStateFault :: AsError a => Getting (First ServiceError) a ServiceError

The VPC network is in an invalid state.

_SubnetInUse :: AsError a => Getting (First ServiceError) a ServiceError

The requested subnet is being used by another subnet group.

_ServiceQuotaExceededException :: AsError a => Getting (First ServiceError) a ServiceError

You have reached the maximum number of x509 certificates that can be created for encrypted clusters in a 30 day period. Contact AWS customer support to discuss options for continuing to create encrypted clusters.

_ClusterQuotaForCustomerExceededFault :: AsError a => Getting (First ServiceError) a ServiceError

You have attempted to exceed the maximum number of DAX clusters for your AWS account.

_SubnetGroupNotFoundFault :: AsError a => Getting (First ServiceError) a ServiceError

The requested subnet group name does not refer to an existing subnet group.

_SubnetGroupAlreadyExistsFault :: AsError a => Getting (First ServiceError) a ServiceError

The specified subnet group already exists.

_NodeQuotaForCustomerExceededFault :: AsError a => Getting (First ServiceError) a ServiceError

You have attempted to exceed the maximum number of nodes for your AWS account.

_SubnetGroupQuotaExceededFault :: AsError a => Getting (First ServiceError) a ServiceError

The request cannot be processed because it would exceed the allowed number of subnets in a subnet group.

_ClusterAlreadyExistsFault :: AsError a => Getting (First ServiceError) a ServiceError

You already have a DAX cluster with the given identifier.

_InvalidARNFault :: AsError a => Getting (First ServiceError) a ServiceError

The Amazon Resource Name (ARN) supplied in the request is not valid.

_InvalidParameterCombinationException :: AsError a => Getting (First ServiceError) a ServiceError

Two or more incompatible parameters were specified.

ChangeType

newtype ChangeType

Constructors

ChangeType'

Fields

fromChangeType :: Text

Bundled Patterns

pattern ChangeType_IMMEDIATE :: ChangeType

pattern ChangeType_REQUIRES_REBOOT :: ChangeType

ClusterEndpointEncryptionType

newtype ClusterEndpointEncryptionType

Constructors

ClusterEndpointEncryptionType'

Fields

fromClusterEndpointEncryptionType :: Text

Bundled Patterns

pattern ClusterEndpointEncryptionType_NONE :: ClusterEndpointEncryptionType

pattern ClusterEndpointEncryptionType_TLS :: ClusterEndpointEncryptionType

IsModifiable

newtype IsModifiable

Constructors

IsModifiable'

Fields

fromIsModifiable :: Text

Bundled Patterns

pattern IsModifiable_CONDITIONAL :: IsModifiable

pattern IsModifiable_FALSE :: IsModifiable

pattern IsModifiable_TRUE :: IsModifiable

ParameterType

newtype ParameterType

Constructors

ParameterType'

Fields

fromParameterType :: Text

Bundled Patterns

pattern ParameterType_DEFAULT :: ParameterType

pattern ParameterType_NODE_TYPE_SPECIFIC :: ParameterType

SSEStatus

newtype SSEStatus

Constructors

SSEStatus'

Fields

fromSSEStatus :: Text

Bundled Patterns

pattern SSEStatus_DISABLED :: SSEStatus

pattern SSEStatus_DISABLING :: SSEStatus

pattern SSEStatus_ENABLED :: SSEStatus

pattern SSEStatus_ENABLING :: SSEStatus

SourceType

newtype SourceType

Constructors

SourceType'

Fields

fromSourceType :: Text

Bundled Patterns

pattern SourceType_CLUSTER :: SourceType

pattern SourceType_PARAMETER_GROUP :: SourceType

pattern SourceType_SUBNET_GROUP :: SourceType

Cluster

data Cluster

Contains all of the attributes of a specific DAX cluster.

See: newCluster smart constructor.

Constructors

Cluster'

Fields

status :: Maybe Text

The current status of the cluster.

iamRoleArn :: Maybe Text

A valid Amazon Resource Name (ARN) that identifies an IAM role. At runtime, DAX will assume this role and use the role's permissions to access DynamoDB on your behalf.

clusterArn :: Maybe Text

The Amazon Resource Name (ARN) that uniquely identifies the cluster.

activeNodes :: Maybe Int

The number of nodes in the cluster that are active (i.e., capable of serving requests).

securityGroups :: Maybe [SecurityGroupMembership]

A list of security groups, and the status of each, for the nodes in the cluster.

notificationConfiguration :: Maybe NotificationConfiguration

Describes a notification topic and its status. Notification topics are used for publishing DAX events to subscribers using Amazon Simple Notification Service (SNS).

nodeIdsToRemove :: Maybe [Text]

A list of nodes to be removed from the cluster.

clusterEndpointEncryptionType :: Maybe ClusterEndpointEncryptionType

The type of encryption supported by the cluster's endpoint. Values are:

NONE for no encryption

TLS for Transport Layer Security

totalNodes :: Maybe Int

The total number of nodes in the cluster.

preferredMaintenanceWindow :: Maybe Text

A range of time when maintenance of DAX cluster software will be performed. For example: sun:01:00-sun:09:00. Cluster maintenance normally takes less than 30 minutes, and is performed automatically within the maintenance window.

subnetGroup :: Maybe Text

The subnet group where the DAX cluster is running.

clusterName :: Maybe Text

The name of the DAX cluster.

nodeType :: Maybe Text

The node type for the nodes in the cluster. (All nodes in a DAX cluster are of the same type.)

nodes :: Maybe [Node]

A list of nodes that are currently in the cluster.

clusterDiscoveryEndpoint :: Maybe Endpoint

The endpoint for this DAX cluster, consisting of a DNS name, a port number, and a URL. Applications should use the URL to configure the DAX client to find their cluster.

sSEDescription :: Maybe SSEDescription

The description of the server-side encryption status on the specified DAX cluster.

description :: Maybe Text

The description of the cluster.

parameterGroup :: Maybe ParameterGroupStatus

The parameter group being used by nodes in the cluster.

newCluster :: Cluster

Create a value of Cluster with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:status:Cluster', cluster_status - The current status of the cluster.

$sel:iamRoleArn:Cluster', cluster_iamRoleArn - A valid Amazon Resource Name (ARN) that identifies an IAM role. At runtime, DAX will assume this role and use the role's permissions to access DynamoDB on your behalf.

$sel:clusterArn:Cluster', cluster_clusterArn - The Amazon Resource Name (ARN) that uniquely identifies the cluster.

$sel:activeNodes:Cluster', cluster_activeNodes - The number of nodes in the cluster that are active (i.e., capable of serving requests).

$sel:securityGroups:Cluster', cluster_securityGroups - A list of security groups, and the status of each, for the nodes in the cluster.

$sel:notificationConfiguration:Cluster', cluster_notificationConfiguration - Describes a notification topic and its status. Notification topics are used for publishing DAX events to subscribers using Amazon Simple Notification Service (SNS).

$sel:nodeIdsToRemove:Cluster', cluster_nodeIdsToRemove - A list of nodes to be removed from the cluster.

$sel:clusterEndpointEncryptionType:Cluster', cluster_clusterEndpointEncryptionType - The type of encryption supported by the cluster's endpoint. Values are:

NONE for no encryption

TLS for Transport Layer Security

$sel:totalNodes:Cluster', cluster_totalNodes - The total number of nodes in the cluster.

$sel:preferredMaintenanceWindow:Cluster', cluster_preferredMaintenanceWindow - A range of time when maintenance of DAX cluster software will be performed. For example: sun:01:00-sun:09:00. Cluster maintenance normally takes less than 30 minutes, and is performed automatically within the maintenance window.

$sel:subnetGroup:Cluster', cluster_subnetGroup - The subnet group where the DAX cluster is running.

$sel:clusterName:Cluster', cluster_clusterName - The name of the DAX cluster.

$sel:nodeType:Cluster', cluster_nodeType - The node type for the nodes in the cluster. (All nodes in a DAX cluster are of the same type.)

$sel:nodes:Cluster', cluster_nodes - A list of nodes that are currently in the cluster.

$sel:clusterDiscoveryEndpoint:Cluster', cluster_clusterDiscoveryEndpoint - The endpoint for this DAX cluster, consisting of a DNS name, a port number, and a URL. Applications should use the URL to configure the DAX client to find their cluster.

$sel:sSEDescription:Cluster', cluster_sSEDescription - The description of the server-side encryption status on the specified DAX cluster.

$sel:description:Cluster', cluster_description - The description of the cluster.

$sel:parameterGroup:Cluster', cluster_parameterGroup - The parameter group being used by nodes in the cluster.

cluster_status :: Lens' Cluster (Maybe Text)

The current status of the cluster.

cluster_iamRoleArn :: Lens' Cluster (Maybe Text)

A valid Amazon Resource Name (ARN) that identifies an IAM role. At runtime, DAX will assume this role and use the role's permissions to access DynamoDB on your behalf.

cluster_clusterArn :: Lens' Cluster (Maybe Text)

The Amazon Resource Name (ARN) that uniquely identifies the cluster.

cluster_activeNodes :: Lens' Cluster (Maybe Int)

The number of nodes in the cluster that are active (i.e., capable of serving requests).

cluster_securityGroups :: Lens' Cluster (Maybe [SecurityGroupMembership])

A list of security groups, and the status of each, for the nodes in the cluster.

cluster_notificationConfiguration :: Lens' Cluster (Maybe NotificationConfiguration)

Describes a notification topic and its status. Notification topics are used for publishing DAX events to subscribers using Amazon Simple Notification Service (SNS).

cluster_nodeIdsToRemove :: Lens' Cluster (Maybe [Text])

A list of nodes to be removed from the cluster.

cluster_clusterEndpointEncryptionType :: Lens' Cluster (Maybe ClusterEndpointEncryptionType)

The type of encryption supported by the cluster's endpoint. Values are:

NONE for no encryption

TLS for Transport Layer Security

cluster_totalNodes :: Lens' Cluster (Maybe Int)

The total number of nodes in the cluster.

cluster_preferredMaintenanceWindow :: Lens' Cluster (Maybe Text)

A range of time when maintenance of DAX cluster software will be performed. For example: sun:01:00-sun:09:00. Cluster maintenance normally takes less than 30 minutes, and is performed automatically within the maintenance window.

cluster_subnetGroup :: Lens' Cluster (Maybe Text)

The subnet group where the DAX cluster is running.

cluster_clusterName :: Lens' Cluster (Maybe Text)

The name of the DAX cluster.

cluster_nodeType :: Lens' Cluster (Maybe Text)

The node type for the nodes in the cluster. (All nodes in a DAX cluster are of the same type.)

cluster_nodes :: Lens' Cluster (Maybe [Node])

A list of nodes that are currently in the cluster.

cluster_clusterDiscoveryEndpoint :: Lens' Cluster (Maybe Endpoint)

The endpoint for this DAX cluster, consisting of a DNS name, a port number, and a URL. Applications should use the URL to configure the DAX client to find their cluster.

cluster_sSEDescription :: Lens' Cluster (Maybe SSEDescription)

The description of the server-side encryption status on the specified DAX cluster.

cluster_description :: Lens' Cluster (Maybe Text)

The description of the cluster.

cluster_parameterGroup :: Lens' Cluster (Maybe ParameterGroupStatus)

The parameter group being used by nodes in the cluster.

Endpoint

data Endpoint

Represents the information required for client programs to connect to the endpoint for a DAX cluster.

See: newEndpoint smart constructor.

Constructors

Endpoint'

Fields

url :: Maybe Text

The URL that applications should use to connect to the endpoint. The default ports are 8111 for the "dax" protocol and 9111 for the "daxs" protocol.

address :: Maybe Text

The DNS hostname of the endpoint.

port :: Maybe Int

The port number that applications should use to connect to the endpoint.

newEndpoint :: Endpoint

Create a value of Endpoint with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:url:Endpoint', endpoint_url - The URL that applications should use to connect to the endpoint. The default ports are 8111 for the "dax" protocol and 9111 for the "daxs" protocol.

$sel:address:Endpoint', endpoint_address - The DNS hostname of the endpoint.

$sel:port:Endpoint', endpoint_port - The port number that applications should use to connect to the endpoint.

endpoint_url :: Lens' Endpoint (Maybe Text)

The URL that applications should use to connect to the endpoint. The default ports are 8111 for the "dax" protocol and 9111 for the "daxs" protocol.

endpoint_address :: Lens' Endpoint (Maybe Text)

The DNS hostname of the endpoint.

endpoint_port :: Lens' Endpoint (Maybe Int)

The port number that applications should use to connect to the endpoint.

Event

data Event

Represents a single occurrence of something interesting within the system. Some examples of events are creating a DAX cluster, adding or removing a node, or rebooting a node.

See: newEvent smart constructor.

Constructors

Event'

Fields

sourceName :: Maybe Text

The source of the event. For example, if the event occurred at the node level, the source would be the node ID.

sourceType :: Maybe SourceType

Specifies the origin of this event - a cluster, a parameter group, a node ID, etc.

date :: Maybe POSIX

The date and time when the event occurred.

message :: Maybe Text

A user-defined message associated with the event.

newEvent :: Event

Create a value of Event with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:sourceName:Event', event_sourceName - The source of the event. For example, if the event occurred at the node level, the source would be the node ID.

$sel:sourceType:Event', event_sourceType - Specifies the origin of this event - a cluster, a parameter group, a node ID, etc.

$sel:date:Event', event_date - The date and time when the event occurred.

$sel:message:Event', event_message - A user-defined message associated with the event.

event_sourceName :: Lens' Event (Maybe Text)

The source of the event. For example, if the event occurred at the node level, the source would be the node ID.

event_sourceType :: Lens' Event (Maybe SourceType)

Specifies the origin of this event - a cluster, a parameter group, a node ID, etc.

event_date :: Lens' Event (Maybe UTCTime)

The date and time when the event occurred.

event_message :: Lens' Event (Maybe Text)

A user-defined message associated with the event.

Node

data Node

Represents an individual node within a DAX cluster.

See: newNode smart constructor.

Constructors

Node'

Fields

nodeStatus :: Maybe Text

The current status of the node. For example: available.

parameterGroupStatus :: Maybe Text

The status of the parameter group associated with this node. For example, in-sync.

availabilityZone :: Maybe Text

The Availability Zone (AZ) in which the node has been deployed.

nodeId :: Maybe Text

A system-generated identifier for the node.

endpoint :: Maybe Endpoint

The endpoint for the node, consisting of a DNS name and a port number. Client applications can connect directly to a node endpoint, if desired (as an alternative to allowing DAX client software to intelligently route requests and responses to nodes in the DAX cluster.

nodeCreateTime :: Maybe POSIX

The date and time (in UNIX epoch format) when the node was launched.

newNode :: Node

Create a value of Node with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:nodeStatus:Node', node_nodeStatus - The current status of the node. For example: available.

$sel:parameterGroupStatus:Node', node_parameterGroupStatus - The status of the parameter group associated with this node. For example, in-sync.

$sel:availabilityZone:Node', node_availabilityZone - The Availability Zone (AZ) in which the node has been deployed.

$sel:nodeId:Node', node_nodeId - A system-generated identifier for the node.

$sel:endpoint:Node', node_endpoint - The endpoint for the node, consisting of a DNS name and a port number. Client applications can connect directly to a node endpoint, if desired (as an alternative to allowing DAX client software to intelligently route requests and responses to nodes in the DAX cluster.

$sel:nodeCreateTime:Node', node_nodeCreateTime - The date and time (in UNIX epoch format) when the node was launched.

node_nodeStatus :: Lens' Node (Maybe Text)

The current status of the node. For example: available.

node_parameterGroupStatus :: Lens' Node (Maybe Text)

The status of the parameter group associated with this node. For example, in-sync.

node_availabilityZone :: Lens' Node (Maybe Text)

The Availability Zone (AZ) in which the node has been deployed.

node_nodeId :: Lens' Node (Maybe Text)

A system-generated identifier for the node.

node_endpoint :: Lens' Node (Maybe Endpoint)

The endpoint for the node, consisting of a DNS name and a port number. Client applications can connect directly to a node endpoint, if desired (as an alternative to allowing DAX client software to intelligently route requests and responses to nodes in the DAX cluster.

node_nodeCreateTime :: Lens' Node (Maybe UTCTime)

The date and time (in UNIX epoch format) when the node was launched.

NodeTypeSpecificValue

data NodeTypeSpecificValue

Represents a parameter value that is applicable to a particular node type.

See: newNodeTypeSpecificValue smart constructor.

Constructors

NodeTypeSpecificValue'

Fields

value :: Maybe Text

The parameter value for this node type.

nodeType :: Maybe Text

A node type to which the parameter value applies.

newNodeTypeSpecificValue :: NodeTypeSpecificValue

Create a value of NodeTypeSpecificValue with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:value:NodeTypeSpecificValue', nodeTypeSpecificValue_value - The parameter value for this node type.

$sel:nodeType:NodeTypeSpecificValue', nodeTypeSpecificValue_nodeType - A node type to which the parameter value applies.

nodeTypeSpecificValue_value :: Lens' NodeTypeSpecificValue (Maybe Text)

The parameter value for this node type.

nodeTypeSpecificValue_nodeType :: Lens' NodeTypeSpecificValue (Maybe Text)

A node type to which the parameter value applies.

NotificationConfiguration

data NotificationConfiguration

Describes a notification topic and its status. Notification topics are used for publishing DAX events to subscribers using Amazon Simple Notification Service (SNS).

See: newNotificationConfiguration smart constructor.

Constructors

NotificationConfiguration'

Fields

topicStatus :: Maybe Text

The current state of the topic. A value of “active” means that notifications will be sent to the topic. A value of “inactive” means that notifications will not be sent to the topic.

topicArn :: Maybe Text

The Amazon Resource Name (ARN) that identifies the topic.

newNotificationConfiguration :: NotificationConfiguration

Create a value of NotificationConfiguration with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:topicStatus:NotificationConfiguration', notificationConfiguration_topicStatus - The current state of the topic. A value of “active” means that notifications will be sent to the topic. A value of “inactive” means that notifications will not be sent to the topic.

$sel:topicArn:NotificationConfiguration', notificationConfiguration_topicArn - The Amazon Resource Name (ARN) that identifies the topic.

notificationConfiguration_topicStatus :: Lens' NotificationConfiguration (Maybe Text)

The current state of the topic. A value of “active” means that notifications will be sent to the topic. A value of “inactive” means that notifications will not be sent to the topic.

notificationConfiguration_topicArn :: Lens' NotificationConfiguration (Maybe Text)

The Amazon Resource Name (ARN) that identifies the topic.

Parameter

data Parameter

Describes an individual setting that controls some aspect of DAX behavior.

See: newParameter smart constructor.

Constructors

Parameter'

Fields

parameterValue :: Maybe Text

The value for the parameter.

parameterType :: Maybe ParameterType

Determines whether the parameter can be applied to any nodes, or only nodes of a particular type.

source :: Maybe Text

How the parameter is defined. For example, system denotes a system-defined parameter.

isModifiable :: Maybe IsModifiable

Whether the customer is allowed to modify the parameter.

dataType :: Maybe Text

The data type of the parameter. For example, integer:

nodeTypeSpecificValues :: Maybe [NodeTypeSpecificValue]

A list of node types, and specific parameter values for each node.

allowedValues :: Maybe Text

A range of values within which the parameter can be set.

parameterName :: Maybe Text

The name of the parameter.

description :: Maybe Text

A description of the parameter

changeType :: Maybe ChangeType

The conditions under which changes to this parameter can be applied. For example, requires-reboot indicates that a new value for this parameter will only take effect if a node is rebooted.

newParameter :: Parameter

Create a value of Parameter with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:parameterValue:Parameter', parameter_parameterValue - The value for the parameter.

$sel:parameterType:Parameter', parameter_parameterType - Determines whether the parameter can be applied to any nodes, or only nodes of a particular type.

$sel:source:Parameter', parameter_source - How the parameter is defined. For example, system denotes a system-defined parameter.

$sel:isModifiable:Parameter', parameter_isModifiable - Whether the customer is allowed to modify the parameter.

$sel:dataType:Parameter', parameter_dataType - The data type of the parameter. For example, integer:

$sel:nodeTypeSpecificValues:Parameter', parameter_nodeTypeSpecificValues - A list of node types, and specific parameter values for each node.

$sel:allowedValues:Parameter', parameter_allowedValues - A range of values within which the parameter can be set.

$sel:parameterName:Parameter', parameter_parameterName - The name of the parameter.

$sel:description:Parameter', parameter_description - A description of the parameter

$sel:changeType:Parameter', parameter_changeType - The conditions under which changes to this parameter can be applied. For example, requires-reboot indicates that a new value for this parameter will only take effect if a node is rebooted.

parameter_parameterValue :: Lens' Parameter (Maybe Text)

The value for the parameter.

parameter_parameterType :: Lens' Parameter (Maybe ParameterType)

Determines whether the parameter can be applied to any nodes, or only nodes of a particular type.

parameter_source :: Lens' Parameter (Maybe Text)

How the parameter is defined. For example, system denotes a system-defined parameter.

parameter_isModifiable :: Lens' Parameter (Maybe IsModifiable)

Whether the customer is allowed to modify the parameter.

parameter_dataType :: Lens' Parameter (Maybe Text)

The data type of the parameter. For example, integer:

parameter_nodeTypeSpecificValues :: Lens' Parameter (Maybe [NodeTypeSpecificValue])

A list of node types, and specific parameter values for each node.

parameter_allowedValues :: Lens' Parameter (Maybe Text)

A range of values within which the parameter can be set.

parameter_parameterName :: Lens' Parameter (Maybe Text)

The name of the parameter.

parameter_description :: Lens' Parameter (Maybe Text)

A description of the parameter

parameter_changeType :: Lens' Parameter (Maybe ChangeType)

The conditions under which changes to this parameter can be applied. For example, requires-reboot indicates that a new value for this parameter will only take effect if a node is rebooted.

ParameterGroup

data ParameterGroup

A named set of parameters that are applied to all of the nodes in a DAX cluster.

See: newParameterGroup smart constructor.

Constructors

ParameterGroup'

Fields

description :: Maybe Text

A description of the parameter group.

parameterGroupName :: Maybe Text

The name of the parameter group.

newParameterGroup :: ParameterGroup

Create a value of ParameterGroup with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:description:ParameterGroup', parameterGroup_description - A description of the parameter group.

$sel:parameterGroupName:ParameterGroup', parameterGroup_parameterGroupName - The name of the parameter group.

parameterGroup_description :: Lens' ParameterGroup (Maybe Text)

A description of the parameter group.

parameterGroup_parameterGroupName :: Lens' ParameterGroup (Maybe Text)

The name of the parameter group.

ParameterGroupStatus

data ParameterGroupStatus

The status of a parameter group.

See: newParameterGroupStatus smart constructor.

Constructors

ParameterGroupStatus'

Fields

nodeIdsToReboot :: Maybe [Text]

The node IDs of one or more nodes to be rebooted.

parameterApplyStatus :: Maybe Text

The status of parameter updates.

parameterGroupName :: Maybe Text

The name of the parameter group.

newParameterGroupStatus :: ParameterGroupStatus

Create a value of ParameterGroupStatus with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:nodeIdsToReboot:ParameterGroupStatus', parameterGroupStatus_nodeIdsToReboot - The node IDs of one or more nodes to be rebooted.

$sel:parameterApplyStatus:ParameterGroupStatus', parameterGroupStatus_parameterApplyStatus - The status of parameter updates.

$sel:parameterGroupName:ParameterGroupStatus', parameterGroupStatus_parameterGroupName - The name of the parameter group.

parameterGroupStatus_nodeIdsToReboot :: Lens' ParameterGroupStatus (Maybe [Text])

The node IDs of one or more nodes to be rebooted.

parameterGroupStatus_parameterApplyStatus :: Lens' ParameterGroupStatus (Maybe Text)

The status of parameter updates.

parameterGroupStatus_parameterGroupName :: Lens' ParameterGroupStatus (Maybe Text)

The name of the parameter group.

ParameterNameValue

data ParameterNameValue

An individual DAX parameter.

See: newParameterNameValue smart constructor.

Constructors

ParameterNameValue'

Fields

parameterValue :: Maybe Text

The value of the parameter.

parameterName :: Maybe Text

The name of the parameter.

newParameterNameValue :: ParameterNameValue

Create a value of ParameterNameValue with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:parameterValue:ParameterNameValue', parameterNameValue_parameterValue - The value of the parameter.

$sel:parameterName:ParameterNameValue', parameterNameValue_parameterName - The name of the parameter.

parameterNameValue_parameterValue :: Lens' ParameterNameValue (Maybe Text)

The value of the parameter.

parameterNameValue_parameterName :: Lens' ParameterNameValue (Maybe Text)

The name of the parameter.

SSEDescription

data SSEDescription

The description of the server-side encryption status on the specified DAX cluster.

See: newSSEDescription smart constructor.

Constructors

SSEDescription'

Fields

status :: Maybe SSEStatus

The current state of server-side encryption:

ENABLING - Server-side encryption is being enabled.

ENABLED - Server-side encryption is enabled.

DISABLING - Server-side encryption is being disabled.

DISABLED - Server-side encryption is disabled.

newSSEDescription :: SSEDescription

Create a value of SSEDescription with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:status:SSEDescription', sSEDescription_status - The current state of server-side encryption:

ENABLING - Server-side encryption is being enabled.

ENABLED - Server-side encryption is enabled.

DISABLING - Server-side encryption is being disabled.

DISABLED - Server-side encryption is disabled.

sSEDescription_status :: Lens' SSEDescription (Maybe SSEStatus)

The current state of server-side encryption:

ENABLING - Server-side encryption is being enabled.

ENABLED - Server-side encryption is enabled.

DISABLING - Server-side encryption is being disabled.

DISABLED - Server-side encryption is disabled.

SSESpecification

data SSESpecification

Represents the settings used to enable server-side encryption.

See: newSSESpecification smart constructor.

Constructors

SSESpecification'

Fields

enabled :: Bool

Indicates whether server-side encryption is enabled (true) or disabled (false) on the cluster.

newSSESpecification

Arguments

:: Bool
$sel:enabled:SSESpecification'

-> SSESpecification

Create a value of SSESpecification with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:enabled:SSESpecification', sSESpecification_enabled - Indicates whether server-side encryption is enabled (true) or disabled (false) on the cluster.

sSESpecification_enabled :: Lens' SSESpecification Bool

Indicates whether server-side encryption is enabled (true) or disabled (false) on the cluster.

SecurityGroupMembership

data SecurityGroupMembership

An individual VPC security group and its status.

See: newSecurityGroupMembership smart constructor.

Constructors

SecurityGroupMembership'

Fields

status :: Maybe Text

The status of this security group.

securityGroupIdentifier :: Maybe Text

The unique ID for this security group.

newSecurityGroupMembership :: SecurityGroupMembership

Create a value of SecurityGroupMembership with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:status:SecurityGroupMembership', securityGroupMembership_status - The status of this security group.

$sel:securityGroupIdentifier:SecurityGroupMembership', securityGroupMembership_securityGroupIdentifier - The unique ID for this security group.

securityGroupMembership_status :: Lens' SecurityGroupMembership (Maybe Text)

The status of this security group.

securityGroupMembership_securityGroupIdentifier :: Lens' SecurityGroupMembership (Maybe Text)

The unique ID for this security group.

Subnet

data Subnet

Represents the subnet associated with a DAX cluster. This parameter refers to subnets defined in Amazon Virtual Private Cloud (Amazon VPC) and used with DAX.

See: newSubnet smart constructor.

Constructors

Subnet'

Fields

subnetIdentifier :: Maybe Text

The system-assigned identifier for the subnet.

subnetAvailabilityZone :: Maybe Text

The Availability Zone (AZ) for the subnet.

newSubnet :: Subnet

Create a value of Subnet with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:subnetIdentifier:Subnet', subnet_subnetIdentifier - The system-assigned identifier for the subnet.

$sel:subnetAvailabilityZone:Subnet', subnet_subnetAvailabilityZone - The Availability Zone (AZ) for the subnet.

subnet_subnetIdentifier :: Lens' Subnet (Maybe Text)

The system-assigned identifier for the subnet.

subnet_subnetAvailabilityZone :: Lens' Subnet (Maybe Text)

The Availability Zone (AZ) for the subnet.

SubnetGroup

data SubnetGroup

Represents the output of one of the following actions:

CreateSubnetGroup

ModifySubnetGroup

See: newSubnetGroup smart constructor.

Constructors

SubnetGroup'

Fields

vpcId :: Maybe Text

The Amazon Virtual Private Cloud identifier (VPC ID) of the subnet group.

subnets :: Maybe [Subnet]

A list of subnets associated with the subnet group.

subnetGroupName :: Maybe Text

The name of the subnet group.

description :: Maybe Text

The description of the subnet group.

newSubnetGroup :: SubnetGroup

Create a value of SubnetGroup with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:vpcId:SubnetGroup', subnetGroup_vpcId - The Amazon Virtual Private Cloud identifier (VPC ID) of the subnet group.

$sel:subnets:SubnetGroup', subnetGroup_subnets - A list of subnets associated with the subnet group.

$sel:subnetGroupName:SubnetGroup', subnetGroup_subnetGroupName - The name of the subnet group.

$sel:description:SubnetGroup', subnetGroup_description - The description of the subnet group.

subnetGroup_vpcId :: Lens' SubnetGroup (Maybe Text)

The Amazon Virtual Private Cloud identifier (VPC ID) of the subnet group.

subnetGroup_subnets :: Lens' SubnetGroup (Maybe [Subnet])

A list of subnets associated with the subnet group.

subnetGroup_subnetGroupName :: Lens' SubnetGroup (Maybe Text)

The name of the subnet group.

subnetGroup_description :: Lens' SubnetGroup (Maybe Text)

The description of the subnet group.

Tag

data Tag

A description of a tag. Every tag is a key-value pair. You can add up to 50 tags to a single DAX cluster.

AWS-assigned tag names and values are automatically assigned the aws: prefix, which the user cannot assign. AWS-assigned tag names do not count towards the tag limit of 50. User-assigned tag names have the prefix user:.

You cannot backdate the application of a tag.

See: newTag smart constructor.

Constructors

Tag'

Fields

value :: Maybe Text

The value of the tag. Tag values are case-sensitive and can be null.

key :: Maybe Text

The key for the tag. Tag keys are case sensitive. Every DAX cluster can only have one tag with the same key. If you try to add an existing tag (same key), the existing tag value will be updated to the new value.

newTag :: Tag

Create a value of Tag with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:value:Tag', tag_value - The value of the tag. Tag values are case-sensitive and can be null.

$sel:key:Tag', tag_key - The key for the tag. Tag keys are case sensitive. Every DAX cluster can only have one tag with the same key. If you try to add an existing tag (same key), the existing tag value will be updated to the new value.

tag_value :: Lens' Tag (Maybe Text)

The value of the tag. Tag values are case-sensitive and can be null.

tag_key :: Lens' Tag (Maybe Text)

The key for the tag. Tag keys are case sensitive. Every DAX cluster can only have one tag with the same key. If you try to add an existing tag (same key), the existing tag value will be updated to the new value.