Amazonka.DAX.Types.Cluster

Description

Documentation

data Cluster

Contains all of the attributes of a specific DAX cluster.

See: newCluster smart constructor.

Constructors

Cluster'

Fields

status :: Maybe Text

The current status of the cluster.

iamRoleArn :: Maybe Text

A valid Amazon Resource Name (ARN) that identifies an IAM role. At runtime, DAX will assume this role and use the role's permissions to access DynamoDB on your behalf.

clusterArn :: Maybe Text

The Amazon Resource Name (ARN) that uniquely identifies the cluster.

activeNodes :: Maybe Int

The number of nodes in the cluster that are active (i.e., capable of serving requests).

securityGroups :: Maybe [SecurityGroupMembership]

A list of security groups, and the status of each, for the nodes in the cluster.

notificationConfiguration :: Maybe NotificationConfiguration

Describes a notification topic and its status. Notification topics are used for publishing DAX events to subscribers using Amazon Simple Notification Service (SNS).

nodeIdsToRemove :: Maybe [Text]

A list of nodes to be removed from the cluster.

clusterEndpointEncryptionType :: Maybe ClusterEndpointEncryptionType

The type of encryption supported by the cluster's endpoint. Values are:

NONE for no encryption

TLS for Transport Layer Security

totalNodes :: Maybe Int

The total number of nodes in the cluster.

preferredMaintenanceWindow :: Maybe Text

A range of time when maintenance of DAX cluster software will be performed. For example: sun:01:00-sun:09:00. Cluster maintenance normally takes less than 30 minutes, and is performed automatically within the maintenance window.

subnetGroup :: Maybe Text

The subnet group where the DAX cluster is running.

clusterName :: Maybe Text

The name of the DAX cluster.

nodeType :: Maybe Text

The node type for the nodes in the cluster. (All nodes in a DAX cluster are of the same type.)

nodes :: Maybe [Node]

A list of nodes that are currently in the cluster.

clusterDiscoveryEndpoint :: Maybe Endpoint

The endpoint for this DAX cluster, consisting of a DNS name, a port number, and a URL. Applications should use the URL to configure the DAX client to find their cluster.

sSEDescription :: Maybe SSEDescription

The description of the server-side encryption status on the specified DAX cluster.

description :: Maybe Text

The description of the cluster.

parameterGroup :: Maybe ParameterGroupStatus

The parameter group being used by nodes in the cluster.

newCluster :: Cluster

Create a value of Cluster with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:status:Cluster', cluster_status - The current status of the cluster.

$sel:iamRoleArn:Cluster', cluster_iamRoleArn - A valid Amazon Resource Name (ARN) that identifies an IAM role. At runtime, DAX will assume this role and use the role's permissions to access DynamoDB on your behalf.

$sel:clusterArn:Cluster', cluster_clusterArn - The Amazon Resource Name (ARN) that uniquely identifies the cluster.

$sel:activeNodes:Cluster', cluster_activeNodes - The number of nodes in the cluster that are active (i.e., capable of serving requests).

$sel:securityGroups:Cluster', cluster_securityGroups - A list of security groups, and the status of each, for the nodes in the cluster.

$sel:notificationConfiguration:Cluster', cluster_notificationConfiguration - Describes a notification topic and its status. Notification topics are used for publishing DAX events to subscribers using Amazon Simple Notification Service (SNS).

$sel:nodeIdsToRemove:Cluster', cluster_nodeIdsToRemove - A list of nodes to be removed from the cluster.

$sel:clusterEndpointEncryptionType:Cluster', cluster_clusterEndpointEncryptionType - The type of encryption supported by the cluster's endpoint. Values are:

NONE for no encryption

TLS for Transport Layer Security

$sel:totalNodes:Cluster', cluster_totalNodes - The total number of nodes in the cluster.

$sel:preferredMaintenanceWindow:Cluster', cluster_preferredMaintenanceWindow - A range of time when maintenance of DAX cluster software will be performed. For example: sun:01:00-sun:09:00. Cluster maintenance normally takes less than 30 minutes, and is performed automatically within the maintenance window.

$sel:subnetGroup:Cluster', cluster_subnetGroup - The subnet group where the DAX cluster is running.

$sel:clusterName:Cluster', cluster_clusterName - The name of the DAX cluster.

$sel:nodeType:Cluster', cluster_nodeType - The node type for the nodes in the cluster. (All nodes in a DAX cluster are of the same type.)

$sel:nodes:Cluster', cluster_nodes - A list of nodes that are currently in the cluster.

$sel:clusterDiscoveryEndpoint:Cluster', cluster_clusterDiscoveryEndpoint - The endpoint for this DAX cluster, consisting of a DNS name, a port number, and a URL. Applications should use the URL to configure the DAX client to find their cluster.

$sel:sSEDescription:Cluster', cluster_sSEDescription - The description of the server-side encryption status on the specified DAX cluster.

$sel:description:Cluster', cluster_description - The description of the cluster.

$sel:parameterGroup:Cluster', cluster_parameterGroup - The parameter group being used by nodes in the cluster.

cluster_status :: Lens' Cluster (Maybe Text)

The current status of the cluster.

cluster_iamRoleArn :: Lens' Cluster (Maybe Text)

A valid Amazon Resource Name (ARN) that identifies an IAM role. At runtime, DAX will assume this role and use the role's permissions to access DynamoDB on your behalf.

cluster_clusterArn :: Lens' Cluster (Maybe Text)

The Amazon Resource Name (ARN) that uniquely identifies the cluster.

cluster_activeNodes :: Lens' Cluster (Maybe Int)

The number of nodes in the cluster that are active (i.e., capable of serving requests).

cluster_securityGroups :: Lens' Cluster (Maybe [SecurityGroupMembership])

A list of security groups, and the status of each, for the nodes in the cluster.

cluster_notificationConfiguration :: Lens' Cluster (Maybe NotificationConfiguration)

Describes a notification topic and its status. Notification topics are used for publishing DAX events to subscribers using Amazon Simple Notification Service (SNS).

cluster_nodeIdsToRemove :: Lens' Cluster (Maybe [Text])

A list of nodes to be removed from the cluster.

cluster_clusterEndpointEncryptionType :: Lens' Cluster (Maybe ClusterEndpointEncryptionType)

The type of encryption supported by the cluster's endpoint. Values are:

NONE for no encryption

TLS for Transport Layer Security

cluster_totalNodes :: Lens' Cluster (Maybe Int)

The total number of nodes in the cluster.

cluster_preferredMaintenanceWindow :: Lens' Cluster (Maybe Text)

A range of time when maintenance of DAX cluster software will be performed. For example: sun:01:00-sun:09:00. Cluster maintenance normally takes less than 30 minutes, and is performed automatically within the maintenance window.

cluster_subnetGroup :: Lens' Cluster (Maybe Text)

The subnet group where the DAX cluster is running.

cluster_clusterName :: Lens' Cluster (Maybe Text)

The name of the DAX cluster.

cluster_nodeType :: Lens' Cluster (Maybe Text)

The node type for the nodes in the cluster. (All nodes in a DAX cluster are of the same type.)

cluster_nodes :: Lens' Cluster (Maybe [Node])

A list of nodes that are currently in the cluster.

cluster_clusterDiscoveryEndpoint :: Lens' Cluster (Maybe Endpoint)

The endpoint for this DAX cluster, consisting of a DNS name, a port number, and a URL. Applications should use the URL to configure the DAX client to find their cluster.

cluster_sSEDescription :: Lens' Cluster (Maybe SSEDescription)

The description of the server-side encryption status on the specified DAX cluster.

cluster_description :: Lens' Cluster (Maybe Text)

The description of the cluster.

cluster_parameterGroup :: Lens' Cluster (Maybe ParameterGroupStatus)

The parameter group being used by nodes in the cluster.