Amazonka.DAX.Types.Parameter

Description

Documentation

data Parameter

Describes an individual setting that controls some aspect of DAX behavior.

See: newParameter smart constructor.

Constructors

Parameter'

Fields

parameterValue :: Maybe Text

The value for the parameter.

parameterType :: Maybe ParameterType

Determines whether the parameter can be applied to any nodes, or only nodes of a particular type.

source :: Maybe Text

How the parameter is defined. For example, system denotes a system-defined parameter.

isModifiable :: Maybe IsModifiable

Whether the customer is allowed to modify the parameter.

dataType :: Maybe Text

The data type of the parameter. For example, integer:

nodeTypeSpecificValues :: Maybe [NodeTypeSpecificValue]

A list of node types, and specific parameter values for each node.

allowedValues :: Maybe Text

A range of values within which the parameter can be set.

parameterName :: Maybe Text

The name of the parameter.

description :: Maybe Text

A description of the parameter

changeType :: Maybe ChangeType

The conditions under which changes to this parameter can be applied. For example, requires-reboot indicates that a new value for this parameter will only take effect if a node is rebooted.

newParameter :: Parameter

Create a value of Parameter with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:parameterValue:Parameter', parameter_parameterValue - The value for the parameter.

$sel:parameterType:Parameter', parameter_parameterType - Determines whether the parameter can be applied to any nodes, or only nodes of a particular type.

$sel:source:Parameter', parameter_source - How the parameter is defined. For example, system denotes a system-defined parameter.

$sel:isModifiable:Parameter', parameter_isModifiable - Whether the customer is allowed to modify the parameter.

$sel:dataType:Parameter', parameter_dataType - The data type of the parameter. For example, integer:

$sel:nodeTypeSpecificValues:Parameter', parameter_nodeTypeSpecificValues - A list of node types, and specific parameter values for each node.

$sel:allowedValues:Parameter', parameter_allowedValues - A range of values within which the parameter can be set.

$sel:parameterName:Parameter', parameter_parameterName - The name of the parameter.

$sel:description:Parameter', parameter_description - A description of the parameter

$sel:changeType:Parameter', parameter_changeType - The conditions under which changes to this parameter can be applied. For example, requires-reboot indicates that a new value for this parameter will only take effect if a node is rebooted.

parameter_parameterValue :: Lens' Parameter (Maybe Text)

The value for the parameter.

parameter_parameterType :: Lens' Parameter (Maybe ParameterType)

Determines whether the parameter can be applied to any nodes, or only nodes of a particular type.

parameter_source :: Lens' Parameter (Maybe Text)

How the parameter is defined. For example, system denotes a system-defined parameter.

parameter_isModifiable :: Lens' Parameter (Maybe IsModifiable)

Whether the customer is allowed to modify the parameter.

parameter_dataType :: Lens' Parameter (Maybe Text)

The data type of the parameter. For example, integer:

parameter_nodeTypeSpecificValues :: Lens' Parameter (Maybe [NodeTypeSpecificValue])

A list of node types, and specific parameter values for each node.

parameter_allowedValues :: Lens' Parameter (Maybe Text)

A range of values within which the parameter can be set.

parameter_parameterName :: Lens' Parameter (Maybe Text)

The name of the parameter.

parameter_description :: Lens' Parameter (Maybe Text)

A description of the parameter

parameter_changeType :: Lens' Parameter (Maybe ChangeType)

The conditions under which changes to this parameter can be applied. For example, requires-reboot indicates that a new value for this parameter will only take effect if a node is rebooted.