Amazonka.DAX.UpdateParameterGroup

Description

Modifies the parameters of a parameter group. You can modify up to 20 parameters in a single request by submitting a list parameter name and value pairs.

Creating a Request

data UpdateParameterGroup

See: newUpdateParameterGroup smart constructor.

Constructors

UpdateParameterGroup'

Fields

parameterGroupName :: Text

The name of the parameter group.

parameterNameValues :: [ParameterNameValue]

An array of name-value pairs for the parameters in the group. Each element in the array represents a single parameter.

record-ttl-millis and query-ttl-millis are the only supported parameter names. For more details, see Configuring TTL Settings.

newUpdateParameterGroup

Arguments

:: Text
$sel:parameterGroupName:UpdateParameterGroup'

-> UpdateParameterGroup

Create a value of UpdateParameterGroup with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:parameterGroupName:UpdateParameterGroup', updateParameterGroup_parameterGroupName - The name of the parameter group.

$sel:parameterNameValues:UpdateParameterGroup', updateParameterGroup_parameterNameValues - An array of name-value pairs for the parameters in the group. Each element in the array represents a single parameter.

record-ttl-millis and query-ttl-millis are the only supported parameter names. For more details, see Configuring TTL Settings.

Request Lenses

updateParameterGroup_parameterGroupName :: Lens' UpdateParameterGroup Text

The name of the parameter group.

updateParameterGroup_parameterNameValues :: Lens' UpdateParameterGroup [ParameterNameValue]

An array of name-value pairs for the parameters in the group. Each element in the array represents a single parameter.

record-ttl-millis and query-ttl-millis are the only supported parameter names. For more details, see Configuring TTL Settings.

Destructuring the Response

data UpdateParameterGroupResponse

See: newUpdateParameterGroupResponse smart constructor.

Constructors

UpdateParameterGroupResponse'

Fields

parameterGroup :: Maybe ParameterGroup

The parameter group that has been modified.

httpStatus :: Int

The response's http status code.

newUpdateParameterGroupResponse

Arguments

:: Int
$sel:httpStatus:UpdateParameterGroupResponse'

-> UpdateParameterGroupResponse

Create a value of UpdateParameterGroupResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:parameterGroup:UpdateParameterGroupResponse', updateParameterGroupResponse_parameterGroup - The parameter group that has been modified.

$sel:httpStatus:UpdateParameterGroupResponse', updateParameterGroupResponse_httpStatus - The response's http status code.

Response Lenses

updateParameterGroupResponse_parameterGroup :: Lens' UpdateParameterGroupResponse (Maybe ParameterGroup)

The parameter group that has been modified.

updateParameterGroupResponse_httpStatus :: Lens' UpdateParameterGroupResponse Int

The response's http status code.