Amazonka.DLM.Types

Description

Service Configuration

defaultService :: Service

API version 2018-01-12 of the Amazon Data Lifecycle Manager SDK configuration.

Errors

_InvalidRequestException :: AsError a => Getting (First ServiceError) a ServiceError

Bad request. The request is missing required parameters or has invalid parameters.

_InternalServerException :: AsError a => Getting (First ServiceError) a ServiceError

The service failed in an unexpected way.

_ResourceNotFoundException :: AsError a => Getting (First ServiceError) a ServiceError

A requested resource was not found.

_LimitExceededException :: AsError a => Getting (First ServiceError) a ServiceError

The request failed because a limit was exceeded.

EventSourceValues

newtype EventSourceValues

Constructors

EventSourceValues'

Fields

fromEventSourceValues :: Text

Bundled Patterns

pattern EventSourceValues_MANAGED_CWE :: EventSourceValues

EventTypeValues

newtype EventTypeValues

Constructors

EventTypeValues'

Fields

fromEventTypeValues :: Text

Bundled Patterns

pattern EventTypeValues_ShareSnapshot :: EventTypeValues

GettablePolicyStateValues

newtype GettablePolicyStateValues

Constructors

GettablePolicyStateValues'

Fields

fromGettablePolicyStateValues :: Text

Bundled Patterns

pattern GettablePolicyStateValues_DISABLED :: GettablePolicyStateValues

pattern GettablePolicyStateValues_ENABLED :: GettablePolicyStateValues

pattern GettablePolicyStateValues_ERROR :: GettablePolicyStateValues

IntervalUnitValues

newtype IntervalUnitValues

Constructors

IntervalUnitValues'

Fields

fromIntervalUnitValues :: Text

Bundled Patterns

pattern IntervalUnitValues_HOURS :: IntervalUnitValues

LocationValues

newtype LocationValues

Constructors

LocationValues'

Fields

fromLocationValues :: Text

Bundled Patterns

pattern LocationValues_CLOUD :: LocationValues

pattern LocationValues_OUTPOST_LOCAL :: LocationValues

PolicyTypeValues

newtype PolicyTypeValues

Constructors

PolicyTypeValues'

Fields

fromPolicyTypeValues :: Text

Bundled Patterns

pattern PolicyTypeValues_EBS_SNAPSHOT_MANAGEMENT :: PolicyTypeValues

pattern PolicyTypeValues_EVENT_BASED_POLICY :: PolicyTypeValues

pattern PolicyTypeValues_IMAGE_MANAGEMENT :: PolicyTypeValues

ResourceLocationValues

newtype ResourceLocationValues

Constructors

ResourceLocationValues'

Fields

fromResourceLocationValues :: Text

Bundled Patterns

pattern ResourceLocationValues_CLOUD :: ResourceLocationValues

pattern ResourceLocationValues_OUTPOST :: ResourceLocationValues

ResourceTypeValues

newtype ResourceTypeValues

Constructors

ResourceTypeValues'

Fields

fromResourceTypeValues :: Text

Bundled Patterns

pattern ResourceTypeValues_INSTANCE :: ResourceTypeValues

pattern ResourceTypeValues_VOLUME :: ResourceTypeValues

RetentionIntervalUnitValues

newtype RetentionIntervalUnitValues

Constructors

RetentionIntervalUnitValues'

Fields

fromRetentionIntervalUnitValues :: Text

Bundled Patterns

pattern RetentionIntervalUnitValues_DAYS :: RetentionIntervalUnitValues

pattern RetentionIntervalUnitValues_MONTHS :: RetentionIntervalUnitValues

pattern RetentionIntervalUnitValues_WEEKS :: RetentionIntervalUnitValues

pattern RetentionIntervalUnitValues_YEARS :: RetentionIntervalUnitValues

SettablePolicyStateValues

newtype SettablePolicyStateValues

Constructors

SettablePolicyStateValues'

Fields

fromSettablePolicyStateValues :: Text

Bundled Patterns

pattern SettablePolicyStateValues_DISABLED :: SettablePolicyStateValues

pattern SettablePolicyStateValues_ENABLED :: SettablePolicyStateValues

Action

data Action

Specifies an action for an event-based policy.

See: newAction smart constructor.

Constructors

Action'

Fields

name :: Text

A descriptive name for the action.

crossRegionCopy :: [CrossRegionCopyAction]

The rule for copying shared snapshots across Regions.

newAction

Arguments

:: Text
$sel:name:Action'

-> Action

Create a value of Action with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:name:Action', action_name - A descriptive name for the action.

$sel:crossRegionCopy:Action', action_crossRegionCopy - The rule for copying shared snapshots across Regions.

action_name :: Lens' Action Text

A descriptive name for the action.

action_crossRegionCopy :: Lens' Action [CrossRegionCopyAction]

The rule for copying shared snapshots across Regions.

CreateRule

data CreateRule

Specifies when to create snapshots of EBS volumes.

You must specify either a Cron expression or an interval, interval unit, and start time. You cannot specify both.

See: newCreateRule smart constructor.

Constructors

CreateRule'

Fields

location :: Maybe LocationValues

Specifies the destination for snapshots created by the policy. To create snapshots in the same Region as the source resource, specify CLOUD. To create snapshots on the same Outpost as the source resource, specify OUTPOST_LOCAL. If you omit this parameter, CLOUD is used by default.

If the policy targets resources in an Amazon Web Services Region, then you must create snapshots in the same Region as the source resource.

If the policy targets resources on an Outpost, then you can create snapshots on the same Outpost as the source resource, or in the Region of that Outpost.

interval :: Maybe Natural

The interval between snapshots. The supported values are 1, 2, 3, 4, 6, 8, 12, and 24.

cronExpression :: Maybe Text

The schedule, as a Cron expression. The schedule interval must be between 1 hour and 1 year. For more information, see Cron expressions in the Amazon CloudWatch User Guide.

times :: Maybe [Text]

The time, in UTC, to start the operation. The supported format is hh:mm.

The operation occurs within a one-hour window following the specified time. If you do not specify a time, Amazon DLM selects a time within the next 24 hours.

intervalUnit :: Maybe IntervalUnitValues

The interval unit.

newCreateRule :: CreateRule

Create a value of CreateRule with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:location:CreateRule', createRule_location - Specifies the destination for snapshots created by the policy. To create snapshots in the same Region as the source resource, specify CLOUD. To create snapshots on the same Outpost as the source resource, specify OUTPOST_LOCAL. If you omit this parameter, CLOUD is used by default.

If the policy targets resources in an Amazon Web Services Region, then you must create snapshots in the same Region as the source resource.

If the policy targets resources on an Outpost, then you can create snapshots on the same Outpost as the source resource, or in the Region of that Outpost.

$sel:interval:CreateRule', createRule_interval - The interval between snapshots. The supported values are 1, 2, 3, 4, 6, 8, 12, and 24.

$sel:cronExpression:CreateRule', createRule_cronExpression - The schedule, as a Cron expression. The schedule interval must be between 1 hour and 1 year. For more information, see Cron expressions in the Amazon CloudWatch User Guide.

$sel:times:CreateRule', createRule_times - The time, in UTC, to start the operation. The supported format is hh:mm.

The operation occurs within a one-hour window following the specified time. If you do not specify a time, Amazon DLM selects a time within the next 24 hours.

$sel:intervalUnit:CreateRule', createRule_intervalUnit - The interval unit.

createRule_location :: Lens' CreateRule (Maybe LocationValues)

Specifies the destination for snapshots created by the policy. To create snapshots in the same Region as the source resource, specify CLOUD. To create snapshots on the same Outpost as the source resource, specify OUTPOST_LOCAL. If you omit this parameter, CLOUD is used by default.

If the policy targets resources in an Amazon Web Services Region, then you must create snapshots in the same Region as the source resource.

If the policy targets resources on an Outpost, then you can create snapshots on the same Outpost as the source resource, or in the Region of that Outpost.

createRule_interval :: Lens' CreateRule (Maybe Natural)

The interval between snapshots. The supported values are 1, 2, 3, 4, 6, 8, 12, and 24.

createRule_cronExpression :: Lens' CreateRule (Maybe Text)

The schedule, as a Cron expression. The schedule interval must be between 1 hour and 1 year. For more information, see Cron expressions in the Amazon CloudWatch User Guide.

createRule_times :: Lens' CreateRule (Maybe [Text])

The time, in UTC, to start the operation. The supported format is hh:mm.

The operation occurs within a one-hour window following the specified time. If you do not specify a time, Amazon DLM selects a time within the next 24 hours.

createRule_intervalUnit :: Lens' CreateRule (Maybe IntervalUnitValues)

The interval unit.

CrossRegionCopyAction

data CrossRegionCopyAction

Specifies a rule for copying shared snapshots across Regions.

See: newCrossRegionCopyAction smart constructor.

Constructors

CrossRegionCopyAction'

Fields

retainRule :: Maybe CrossRegionCopyRetainRule

target :: Text

The target Region.

encryptionConfiguration :: EncryptionConfiguration

The encryption settings for the copied snapshot.

newCrossRegionCopyAction

Arguments

:: Text
$sel:target:CrossRegionCopyAction'

-> EncryptionConfiguration
$sel:encryptionConfiguration:CrossRegionCopyAction'

-> CrossRegionCopyAction

Create a value of CrossRegionCopyAction with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:retainRule:CrossRegionCopyAction', crossRegionCopyAction_retainRule - Undocumented member.

$sel:target:CrossRegionCopyAction', crossRegionCopyAction_target - The target Region.

$sel:encryptionConfiguration:CrossRegionCopyAction', crossRegionCopyAction_encryptionConfiguration - The encryption settings for the copied snapshot.

crossRegionCopyAction_retainRule :: Lens' CrossRegionCopyAction (Maybe CrossRegionCopyRetainRule)

Undocumented member.

crossRegionCopyAction_target :: Lens' CrossRegionCopyAction Text

The target Region.

crossRegionCopyAction_encryptionConfiguration :: Lens' CrossRegionCopyAction EncryptionConfiguration

The encryption settings for the copied snapshot.

CrossRegionCopyDeprecateRule

data CrossRegionCopyDeprecateRule

Specifies an AMI deprecation rule for cross-Region AMI copies created by a cross-Region copy rule.

See: newCrossRegionCopyDeprecateRule smart constructor.

Constructors

CrossRegionCopyDeprecateRule'

Fields

interval :: Maybe Natural

The period after which to deprecate the cross-Region AMI copies. The period must be less than or equal to the cross-Region AMI copy retention period, and it can't be greater than 10 years. This is equivalent to 120 months, 520 weeks, or 3650 days.

intervalUnit :: Maybe RetentionIntervalUnitValues

The unit of time in which to measure the Interval.

newCrossRegionCopyDeprecateRule :: CrossRegionCopyDeprecateRule

Create a value of CrossRegionCopyDeprecateRule with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:interval:CrossRegionCopyDeprecateRule', crossRegionCopyDeprecateRule_interval - The period after which to deprecate the cross-Region AMI copies. The period must be less than or equal to the cross-Region AMI copy retention period, and it can't be greater than 10 years. This is equivalent to 120 months, 520 weeks, or 3650 days.

$sel:intervalUnit:CrossRegionCopyDeprecateRule', crossRegionCopyDeprecateRule_intervalUnit - The unit of time in which to measure the Interval.

crossRegionCopyDeprecateRule_interval :: Lens' CrossRegionCopyDeprecateRule (Maybe Natural)

The period after which to deprecate the cross-Region AMI copies. The period must be less than or equal to the cross-Region AMI copy retention period, and it can't be greater than 10 years. This is equivalent to 120 months, 520 weeks, or 3650 days.

crossRegionCopyDeprecateRule_intervalUnit :: Lens' CrossRegionCopyDeprecateRule (Maybe RetentionIntervalUnitValues)

The unit of time in which to measure the Interval.

CrossRegionCopyRetainRule

data CrossRegionCopyRetainRule

Specifies the retention rule for cross-Region snapshot copies.

See: newCrossRegionCopyRetainRule smart constructor.

Constructors

CrossRegionCopyRetainRule'

Fields

interval :: Maybe Natural

The amount of time to retain each snapshot. The maximum is 100 years. This is equivalent to 1200 months, 5200 weeks, or 36500 days.

intervalUnit :: Maybe RetentionIntervalUnitValues

The unit of time for time-based retention.

newCrossRegionCopyRetainRule :: CrossRegionCopyRetainRule

Create a value of CrossRegionCopyRetainRule with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:interval:CrossRegionCopyRetainRule', crossRegionCopyRetainRule_interval - The amount of time to retain each snapshot. The maximum is 100 years. This is equivalent to 1200 months, 5200 weeks, or 36500 days.

$sel:intervalUnit:CrossRegionCopyRetainRule', crossRegionCopyRetainRule_intervalUnit - The unit of time for time-based retention.

crossRegionCopyRetainRule_interval :: Lens' CrossRegionCopyRetainRule (Maybe Natural)

The amount of time to retain each snapshot. The maximum is 100 years. This is equivalent to 1200 months, 5200 weeks, or 36500 days.

crossRegionCopyRetainRule_intervalUnit :: Lens' CrossRegionCopyRetainRule (Maybe RetentionIntervalUnitValues)

The unit of time for time-based retention.

CrossRegionCopyRule

data CrossRegionCopyRule

Specifies a rule for cross-Region snapshot copies.

See: newCrossRegionCopyRule smart constructor.

Constructors

CrossRegionCopyRule'

Fields

deprecateRule :: Maybe CrossRegionCopyDeprecateRule

The AMI deprecation rule for cross-Region AMI copies created by the rule.

targetRegion :: Maybe Text

Avoid using this parameter when creating new policies. Instead, use Target to specify a target Region or a target Outpost for snapshot copies.

For policies created before the Target parameter was introduced, this parameter indicates the target Region for snapshot copies.

copyTags :: Maybe Bool

Indicates whether to copy all user-defined tags from the source snapshot to the cross-Region snapshot copy.

cmkArn :: Maybe Text

The Amazon Resource Name (ARN) of the KMS key to use for EBS encryption. If this parameter is not specified, the default KMS key for the account is used.

retainRule :: Maybe CrossRegionCopyRetainRule

The retention rule that indicates how long snapshot copies are to be retained in the destination Region.

target :: Maybe Text

The target Region or the Amazon Resource Name (ARN) of the target Outpost for the snapshot copies.

Use this parameter instead of TargetRegion. Do not specify both.

encrypted :: Bool

To encrypt a copy of an unencrypted snapshot if encryption by default is not enabled, enable encryption using this parameter. Copies of encrypted snapshots are encrypted, even if this parameter is false or if encryption by default is not enabled.

newCrossRegionCopyRule

Arguments

:: Bool
$sel:encrypted:CrossRegionCopyRule'

-> CrossRegionCopyRule

Create a value of CrossRegionCopyRule with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:deprecateRule:CrossRegionCopyRule', crossRegionCopyRule_deprecateRule - The AMI deprecation rule for cross-Region AMI copies created by the rule.

$sel:targetRegion:CrossRegionCopyRule', crossRegionCopyRule_targetRegion - Avoid using this parameter when creating new policies. Instead, use Target to specify a target Region or a target Outpost for snapshot copies.

For policies created before the Target parameter was introduced, this parameter indicates the target Region for snapshot copies.

$sel:copyTags:CrossRegionCopyRule', crossRegionCopyRule_copyTags - Indicates whether to copy all user-defined tags from the source snapshot to the cross-Region snapshot copy.

$sel:cmkArn:CrossRegionCopyRule', crossRegionCopyRule_cmkArn - The Amazon Resource Name (ARN) of the KMS key to use for EBS encryption. If this parameter is not specified, the default KMS key for the account is used.

$sel:retainRule:CrossRegionCopyRule', crossRegionCopyRule_retainRule - The retention rule that indicates how long snapshot copies are to be retained in the destination Region.

$sel:target:CrossRegionCopyRule', crossRegionCopyRule_target - The target Region or the Amazon Resource Name (ARN) of the target Outpost for the snapshot copies.

Use this parameter instead of TargetRegion. Do not specify both.

$sel:encrypted:CrossRegionCopyRule', crossRegionCopyRule_encrypted - To encrypt a copy of an unencrypted snapshot if encryption by default is not enabled, enable encryption using this parameter. Copies of encrypted snapshots are encrypted, even if this parameter is false or if encryption by default is not enabled.

crossRegionCopyRule_deprecateRule :: Lens' CrossRegionCopyRule (Maybe CrossRegionCopyDeprecateRule)

The AMI deprecation rule for cross-Region AMI copies created by the rule.

crossRegionCopyRule_targetRegion :: Lens' CrossRegionCopyRule (Maybe Text)

Avoid using this parameter when creating new policies. Instead, use Target to specify a target Region or a target Outpost for snapshot copies.

For policies created before the Target parameter was introduced, this parameter indicates the target Region for snapshot copies.

crossRegionCopyRule_copyTags :: Lens' CrossRegionCopyRule (Maybe Bool)

Indicates whether to copy all user-defined tags from the source snapshot to the cross-Region snapshot copy.

crossRegionCopyRule_cmkArn :: Lens' CrossRegionCopyRule (Maybe Text)

The Amazon Resource Name (ARN) of the KMS key to use for EBS encryption. If this parameter is not specified, the default KMS key for the account is used.

crossRegionCopyRule_retainRule :: Lens' CrossRegionCopyRule (Maybe CrossRegionCopyRetainRule)

The retention rule that indicates how long snapshot copies are to be retained in the destination Region.

crossRegionCopyRule_target :: Lens' CrossRegionCopyRule (Maybe Text)

The target Region or the Amazon Resource Name (ARN) of the target Outpost for the snapshot copies.

Use this parameter instead of TargetRegion. Do not specify both.

crossRegionCopyRule_encrypted :: Lens' CrossRegionCopyRule Bool

To encrypt a copy of an unencrypted snapshot if encryption by default is not enabled, enable encryption using this parameter. Copies of encrypted snapshots are encrypted, even if this parameter is false or if encryption by default is not enabled.

DeprecateRule

data DeprecateRule

Specifies an AMI deprecation rule for a schedule.

See: newDeprecateRule smart constructor.

Constructors

DeprecateRule'

Fields

count :: Maybe Natural

If the schedule has a count-based retention rule, this parameter specifies the number of oldest AMIs to deprecate. The count must be less than or equal to the schedule's retention count, and it can't be greater than 1000.

interval :: Maybe Natural

If the schedule has an age-based retention rule, this parameter specifies the period after which to deprecate AMIs created by the schedule. The period must be less than or equal to the schedule's retention period, and it can't be greater than 10 years. This is equivalent to 120 months, 520 weeks, or 3650 days.

intervalUnit :: Maybe RetentionIntervalUnitValues

The unit of time in which to measure the Interval.

newDeprecateRule :: DeprecateRule

Create a value of DeprecateRule with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:count:DeprecateRule', deprecateRule_count - If the schedule has a count-based retention rule, this parameter specifies the number of oldest AMIs to deprecate. The count must be less than or equal to the schedule's retention count, and it can't be greater than 1000.

$sel:interval:DeprecateRule', deprecateRule_interval - If the schedule has an age-based retention rule, this parameter specifies the period after which to deprecate AMIs created by the schedule. The period must be less than or equal to the schedule's retention period, and it can't be greater than 10 years. This is equivalent to 120 months, 520 weeks, or 3650 days.

$sel:intervalUnit:DeprecateRule', deprecateRule_intervalUnit - The unit of time in which to measure the Interval.

deprecateRule_count :: Lens' DeprecateRule (Maybe Natural)

If the schedule has a count-based retention rule, this parameter specifies the number of oldest AMIs to deprecate. The count must be less than or equal to the schedule's retention count, and it can't be greater than 1000.

deprecateRule_interval :: Lens' DeprecateRule (Maybe Natural)

If the schedule has an age-based retention rule, this parameter specifies the period after which to deprecate AMIs created by the schedule. The period must be less than or equal to the schedule's retention period, and it can't be greater than 10 years. This is equivalent to 120 months, 520 weeks, or 3650 days.

deprecateRule_intervalUnit :: Lens' DeprecateRule (Maybe RetentionIntervalUnitValues)

The unit of time in which to measure the Interval.

EncryptionConfiguration

data EncryptionConfiguration

Specifies the encryption settings for shared snapshots that are copied across Regions.

See: newEncryptionConfiguration smart constructor.

Constructors

EncryptionConfiguration'

Fields

cmkArn :: Maybe Text

The Amazon Resource Name (ARN) of the KMS key to use for EBS encryption. If this parameter is not specified, the default KMS key for the account is used.

encrypted :: Bool

To encrypt a copy of an unencrypted snapshot when encryption by default is not enabled, enable encryption using this parameter. Copies of encrypted snapshots are encrypted, even if this parameter is false or when encryption by default is not enabled.

newEncryptionConfiguration

Arguments

:: Bool
$sel:encrypted:EncryptionConfiguration'

-> EncryptionConfiguration

Create a value of EncryptionConfiguration with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:cmkArn:EncryptionConfiguration', encryptionConfiguration_cmkArn - The Amazon Resource Name (ARN) of the KMS key to use for EBS encryption. If this parameter is not specified, the default KMS key for the account is used.

$sel:encrypted:EncryptionConfiguration', encryptionConfiguration_encrypted - To encrypt a copy of an unencrypted snapshot when encryption by default is not enabled, enable encryption using this parameter. Copies of encrypted snapshots are encrypted, even if this parameter is false or when encryption by default is not enabled.

encryptionConfiguration_cmkArn :: Lens' EncryptionConfiguration (Maybe Text)

The Amazon Resource Name (ARN) of the KMS key to use for EBS encryption. If this parameter is not specified, the default KMS key for the account is used.

encryptionConfiguration_encrypted :: Lens' EncryptionConfiguration Bool

To encrypt a copy of an unencrypted snapshot when encryption by default is not enabled, enable encryption using this parameter. Copies of encrypted snapshots are encrypted, even if this parameter is false or when encryption by default is not enabled.

EventParameters

data EventParameters

Specifies an event that triggers an event-based policy.

See: newEventParameters smart constructor.

Constructors

EventParameters'

Fields

eventType :: EventTypeValues

The type of event. Currently, only snapshot sharing events are supported.

snapshotOwner :: [Text]

The IDs of the Amazon Web Services accounts that can trigger policy by sharing snapshots with your account. The policy only runs if one of the specified Amazon Web Services accounts shares a snapshot with your account.

descriptionRegex :: Text

The snapshot description that can trigger the policy. The description pattern is specified using a regular expression. The policy runs only if a snapshot with a description that matches the specified pattern is shared with your account.

For example, specifying ^.*Created for policy: policy-1234567890abcdef0.*$ configures the policy to run only if snapshots created by policy policy-1234567890abcdef0 are shared with your account.

newEventParameters

Arguments

:: EventTypeValues
$sel:eventType:EventParameters'

-> Text
$sel:descriptionRegex:EventParameters'

-> EventParameters

Create a value of EventParameters with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:eventType:EventParameters', eventParameters_eventType - The type of event. Currently, only snapshot sharing events are supported.

$sel:snapshotOwner:EventParameters', eventParameters_snapshotOwner - The IDs of the Amazon Web Services accounts that can trigger policy by sharing snapshots with your account. The policy only runs if one of the specified Amazon Web Services accounts shares a snapshot with your account.

$sel:descriptionRegex:EventParameters', eventParameters_descriptionRegex - The snapshot description that can trigger the policy. The description pattern is specified using a regular expression. The policy runs only if a snapshot with a description that matches the specified pattern is shared with your account.

For example, specifying ^.*Created for policy: policy-1234567890abcdef0.*$ configures the policy to run only if snapshots created by policy policy-1234567890abcdef0 are shared with your account.

eventParameters_eventType :: Lens' EventParameters EventTypeValues

The type of event. Currently, only snapshot sharing events are supported.

eventParameters_snapshotOwner :: Lens' EventParameters [Text]

The IDs of the Amazon Web Services accounts that can trigger policy by sharing snapshots with your account. The policy only runs if one of the specified Amazon Web Services accounts shares a snapshot with your account.

eventParameters_descriptionRegex :: Lens' EventParameters Text

The snapshot description that can trigger the policy. The description pattern is specified using a regular expression. The policy runs only if a snapshot with a description that matches the specified pattern is shared with your account.

For example, specifying ^.*Created for policy: policy-1234567890abcdef0.*$ configures the policy to run only if snapshots created by policy policy-1234567890abcdef0 are shared with your account.

EventSource

data EventSource

Specifies an event that triggers an event-based policy.

See: newEventSource smart constructor.

Constructors

EventSource'

Fields

parameters :: Maybe EventParameters

Information about the event.

type' :: EventSourceValues

The source of the event. Currently only managed CloudWatch Events rules are supported.

newEventSource

Arguments

:: EventSourceValues
$sel:type':EventSource'

-> EventSource

Create a value of EventSource with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:parameters:EventSource', eventSource_parameters - Information about the event.

$sel:type':EventSource', eventSource_type - The source of the event. Currently only managed CloudWatch Events rules are supported.

eventSource_parameters :: Lens' EventSource (Maybe EventParameters)

Information about the event.

eventSource_type :: Lens' EventSource EventSourceValues

The source of the event. Currently only managed CloudWatch Events rules are supported.

FastRestoreRule

data FastRestoreRule

Specifies a rule for enabling fast snapshot restore. You can enable fast snapshot restore based on either a count or a time interval.

See: newFastRestoreRule smart constructor.

Constructors

FastRestoreRule'

Fields

count :: Maybe Natural

The number of snapshots to be enabled with fast snapshot restore.

interval :: Maybe Natural

The amount of time to enable fast snapshot restore. The maximum is 100 years. This is equivalent to 1200 months, 5200 weeks, or 36500 days.

intervalUnit :: Maybe RetentionIntervalUnitValues

The unit of time for enabling fast snapshot restore.

availabilityZones :: NonEmpty Text

The Availability Zones in which to enable fast snapshot restore.

newFastRestoreRule

Arguments

:: NonEmpty Text
$sel:availabilityZones:FastRestoreRule'

-> FastRestoreRule

Create a value of FastRestoreRule with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:count:FastRestoreRule', fastRestoreRule_count - The number of snapshots to be enabled with fast snapshot restore.

$sel:interval:FastRestoreRule', fastRestoreRule_interval - The amount of time to enable fast snapshot restore. The maximum is 100 years. This is equivalent to 1200 months, 5200 weeks, or 36500 days.

$sel:intervalUnit:FastRestoreRule', fastRestoreRule_intervalUnit - The unit of time for enabling fast snapshot restore.

$sel:availabilityZones:FastRestoreRule', fastRestoreRule_availabilityZones - The Availability Zones in which to enable fast snapshot restore.

fastRestoreRule_count :: Lens' FastRestoreRule (Maybe Natural)

The number of snapshots to be enabled with fast snapshot restore.

fastRestoreRule_interval :: Lens' FastRestoreRule (Maybe Natural)

The amount of time to enable fast snapshot restore. The maximum is 100 years. This is equivalent to 1200 months, 5200 weeks, or 36500 days.

fastRestoreRule_intervalUnit :: Lens' FastRestoreRule (Maybe RetentionIntervalUnitValues)

The unit of time for enabling fast snapshot restore.

fastRestoreRule_availabilityZones :: Lens' FastRestoreRule (NonEmpty Text)

The Availability Zones in which to enable fast snapshot restore.

LifecyclePolicy

data LifecyclePolicy

Detailed information about a lifecycle policy.

See: newLifecyclePolicy smart constructor.

Constructors

LifecyclePolicy'

Fields

state :: Maybe GettablePolicyStateValues

The activation state of the lifecycle policy.

policyDetails :: Maybe PolicyDetails

The configuration of the lifecycle policy

policyId :: Maybe Text

The identifier of the lifecycle policy.

executionRoleArn :: Maybe Text

The Amazon Resource Name (ARN) of the IAM role used to run the operations specified by the lifecycle policy.

dateCreated :: Maybe POSIX

The local date and time when the lifecycle policy was created.

statusMessage :: Maybe Text

The description of the status.

dateModified :: Maybe POSIX

The local date and time when the lifecycle policy was last modified.

policyArn :: Maybe Text

The Amazon Resource Name (ARN) of the policy.

description :: Maybe Text

The description of the lifecycle policy.

tags :: Maybe (HashMap Text Text)

The tags.

newLifecyclePolicy :: LifecyclePolicy

Create a value of LifecyclePolicy with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:state:LifecyclePolicy', lifecyclePolicy_state - The activation state of the lifecycle policy.

$sel:policyDetails:LifecyclePolicy', lifecyclePolicy_policyDetails - The configuration of the lifecycle policy

$sel:policyId:LifecyclePolicy', lifecyclePolicy_policyId - The identifier of the lifecycle policy.

$sel:executionRoleArn:LifecyclePolicy', lifecyclePolicy_executionRoleArn - The Amazon Resource Name (ARN) of the IAM role used to run the operations specified by the lifecycle policy.

$sel:dateCreated:LifecyclePolicy', lifecyclePolicy_dateCreated - The local date and time when the lifecycle policy was created.

$sel:statusMessage:LifecyclePolicy', lifecyclePolicy_statusMessage - The description of the status.

$sel:dateModified:LifecyclePolicy', lifecyclePolicy_dateModified - The local date and time when the lifecycle policy was last modified.

$sel:policyArn:LifecyclePolicy', lifecyclePolicy_policyArn - The Amazon Resource Name (ARN) of the policy.

$sel:description:LifecyclePolicy', lifecyclePolicy_description - The description of the lifecycle policy.

$sel:tags:LifecyclePolicy', lifecyclePolicy_tags - The tags.

lifecyclePolicy_state :: Lens' LifecyclePolicy (Maybe GettablePolicyStateValues)

The activation state of the lifecycle policy.

lifecyclePolicy_policyDetails :: Lens' LifecyclePolicy (Maybe PolicyDetails)

The configuration of the lifecycle policy

lifecyclePolicy_policyId :: Lens' LifecyclePolicy (Maybe Text)

The identifier of the lifecycle policy.

lifecyclePolicy_executionRoleArn :: Lens' LifecyclePolicy (Maybe Text)

The Amazon Resource Name (ARN) of the IAM role used to run the operations specified by the lifecycle policy.

lifecyclePolicy_dateCreated :: Lens' LifecyclePolicy (Maybe UTCTime)

The local date and time when the lifecycle policy was created.

lifecyclePolicy_statusMessage :: Lens' LifecyclePolicy (Maybe Text)

The description of the status.

lifecyclePolicy_dateModified :: Lens' LifecyclePolicy (Maybe UTCTime)

The local date and time when the lifecycle policy was last modified.

lifecyclePolicy_policyArn :: Lens' LifecyclePolicy (Maybe Text)

The Amazon Resource Name (ARN) of the policy.

lifecyclePolicy_description :: Lens' LifecyclePolicy (Maybe Text)

The description of the lifecycle policy.

lifecyclePolicy_tags :: Lens' LifecyclePolicy (Maybe (HashMap Text Text))

The tags.

LifecyclePolicySummary

data LifecyclePolicySummary

Summary information about a lifecycle policy.

See: newLifecyclePolicySummary smart constructor.

Constructors

LifecyclePolicySummary'

Fields

state :: Maybe GettablePolicyStateValues

The activation state of the lifecycle policy.

policyId :: Maybe Text

The identifier of the lifecycle policy.

policyType :: Maybe PolicyTypeValues

The type of policy. EBS_SNAPSHOT_MANAGEMENT indicates that the policy manages the lifecycle of Amazon EBS snapshots. IMAGE_MANAGEMENT indicates that the policy manages the lifecycle of EBS-backed AMIs.

description :: Maybe Text

The description of the lifecycle policy.

tags :: Maybe (HashMap Text Text)

The tags.

newLifecyclePolicySummary :: LifecyclePolicySummary

Create a value of LifecyclePolicySummary with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:state:LifecyclePolicySummary', lifecyclePolicySummary_state - The activation state of the lifecycle policy.

$sel:policyId:LifecyclePolicySummary', lifecyclePolicySummary_policyId - The identifier of the lifecycle policy.

$sel:policyType:LifecyclePolicySummary', lifecyclePolicySummary_policyType - The type of policy. EBS_SNAPSHOT_MANAGEMENT indicates that the policy manages the lifecycle of Amazon EBS snapshots. IMAGE_MANAGEMENT indicates that the policy manages the lifecycle of EBS-backed AMIs.

$sel:description:LifecyclePolicySummary', lifecyclePolicySummary_description - The description of the lifecycle policy.

$sel:tags:LifecyclePolicySummary', lifecyclePolicySummary_tags - The tags.

lifecyclePolicySummary_state :: Lens' LifecyclePolicySummary (Maybe GettablePolicyStateValues)

The activation state of the lifecycle policy.

lifecyclePolicySummary_policyId :: Lens' LifecyclePolicySummary (Maybe Text)

The identifier of the lifecycle policy.

lifecyclePolicySummary_policyType :: Lens' LifecyclePolicySummary (Maybe PolicyTypeValues)

The type of policy. EBS_SNAPSHOT_MANAGEMENT indicates that the policy manages the lifecycle of Amazon EBS snapshots. IMAGE_MANAGEMENT indicates that the policy manages the lifecycle of EBS-backed AMIs.

lifecyclePolicySummary_description :: Lens' LifecyclePolicySummary (Maybe Text)

The description of the lifecycle policy.

lifecyclePolicySummary_tags :: Lens' LifecyclePolicySummary (Maybe (HashMap Text Text))

The tags.

Parameters

data Parameters

Specifies optional parameters to add to a policy. The set of valid parameters depends on the combination of policy type and resource type.

See: newParameters smart constructor.

Constructors

Parameters'

Fields

noReboot :: Maybe Bool

Applies to AMI lifecycle policies only. Indicates whether targeted instances are rebooted when the lifecycle policy runs. true indicates that targeted instances are not rebooted when the policy runs. false indicates that target instances are rebooted when the policy runs. The default is true (instances are not rebooted).

excludeBootVolume :: Maybe Bool

EBS Snapshot Management – Instance policies only

Indicates whether to exclude the root volume from snapshots created using CreateSnapshots. The default is false.

newParameters :: Parameters

Create a value of Parameters with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:noReboot:Parameters', parameters_noReboot - Applies to AMI lifecycle policies only. Indicates whether targeted instances are rebooted when the lifecycle policy runs. true indicates that targeted instances are not rebooted when the policy runs. false indicates that target instances are rebooted when the policy runs. The default is true (instances are not rebooted).

$sel:excludeBootVolume:Parameters', parameters_excludeBootVolume - [EBS Snapshot Management – Instance policies only] Indicates whether to exclude the root volume from snapshots created using CreateSnapshots. The default is false.

parameters_noReboot :: Lens' Parameters (Maybe Bool)

Applies to AMI lifecycle policies only. Indicates whether targeted instances are rebooted when the lifecycle policy runs. true indicates that targeted instances are not rebooted when the policy runs. false indicates that target instances are rebooted when the policy runs. The default is true (instances are not rebooted).

parameters_excludeBootVolume :: Lens' Parameters (Maybe Bool)

EBS Snapshot Management – Instance policies only

Indicates whether to exclude the root volume from snapshots created using CreateSnapshots. The default is false.

PolicyDetails

data PolicyDetails

Specifies the configuration of a lifecycle policy.

See: newPolicyDetails smart constructor.

Constructors

PolicyDetails'

Fields

actions :: Maybe (NonEmpty Action)

The actions to be performed when the event-based policy is triggered. You can specify only one action per policy.

This parameter is required for event-based policies only. If you are creating a snapshot or AMI policy, omit this parameter.

targetTags :: Maybe (NonEmpty Tag)

The single tag that identifies targeted resources for this policy.

This parameter is required for snapshot and AMI policies only. If you are creating an event-based policy, omit this parameter.

policyType :: Maybe PolicyTypeValues

The valid target resource types and actions a policy can manage. Specify EBS_SNAPSHOT_MANAGEMENT to create a lifecycle policy that manages the lifecycle of Amazon EBS snapshots. Specify IMAGE_MANAGEMENT to create a lifecycle policy that manages the lifecycle of EBS-backed AMIs. Specify EVENT_BASED_POLICY to create an event-based policy that performs specific actions when a defined event occurs in your Amazon Web Services account.

The default is EBS_SNAPSHOT_MANAGEMENT.

resourceLocations :: Maybe (NonEmpty ResourceLocationValues)

The location of the resources to backup. If the source resources are located in an Amazon Web Services Region, specify CLOUD. If the source resources are located on an Outpost in your account, specify OUTPOST.

If you specify OUTPOST, Amazon Data Lifecycle Manager backs up all resources of the specified type with matching target tags across all of the Outposts in your account.

parameters :: Maybe Parameters

A set of optional parameters for snapshot and AMI lifecycle policies.

This parameter is required for snapshot and AMI policies only. If you are creating an event-based policy, omit this parameter.

schedules :: Maybe (NonEmpty Schedule)

The schedules of policy-defined actions for snapshot and AMI lifecycle policies. A policy can have up to four schedules—one mandatory schedule and up to three optional schedules.

This parameter is required for snapshot and AMI policies only. If you are creating an event-based policy, omit this parameter.

eventSource :: Maybe EventSource

The event that triggers the event-based policy.

This parameter is required for event-based policies only. If you are creating a snapshot or AMI policy, omit this parameter.

resourceTypes :: Maybe (NonEmpty ResourceTypeValues)

The target resource type for snapshot and AMI lifecycle policies. Use VOLUME to create snapshots of individual volumes or use INSTANCE to create multi-volume snapshots from the volumes for an instance.

This parameter is required for snapshot and AMI policies only. If you are creating an event-based policy, omit this parameter.

newPolicyDetails :: PolicyDetails

Create a value of PolicyDetails with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:actions:PolicyDetails', policyDetails_actions - The actions to be performed when the event-based policy is triggered. You can specify only one action per policy.

This parameter is required for event-based policies only. If you are creating a snapshot or AMI policy, omit this parameter.

$sel:targetTags:PolicyDetails', policyDetails_targetTags - The single tag that identifies targeted resources for this policy.

This parameter is required for snapshot and AMI policies only. If you are creating an event-based policy, omit this parameter.

$sel:policyType:PolicyDetails', policyDetails_policyType - The valid target resource types and actions a policy can manage. Specify EBS_SNAPSHOT_MANAGEMENT to create a lifecycle policy that manages the lifecycle of Amazon EBS snapshots. Specify IMAGE_MANAGEMENT to create a lifecycle policy that manages the lifecycle of EBS-backed AMIs. Specify EVENT_BASED_POLICY to create an event-based policy that performs specific actions when a defined event occurs in your Amazon Web Services account.

The default is EBS_SNAPSHOT_MANAGEMENT.

$sel:resourceLocations:PolicyDetails', policyDetails_resourceLocations - The location of the resources to backup. If the source resources are located in an Amazon Web Services Region, specify CLOUD. If the source resources are located on an Outpost in your account, specify OUTPOST.

If you specify OUTPOST, Amazon Data Lifecycle Manager backs up all resources of the specified type with matching target tags across all of the Outposts in your account.

$sel:parameters:PolicyDetails', policyDetails_parameters - A set of optional parameters for snapshot and AMI lifecycle policies.

This parameter is required for snapshot and AMI policies only. If you are creating an event-based policy, omit this parameter.

$sel:schedules:PolicyDetails', policyDetails_schedules - The schedules of policy-defined actions for snapshot and AMI lifecycle policies. A policy can have up to four schedules—one mandatory schedule and up to three optional schedules.

This parameter is required for snapshot and AMI policies only. If you are creating an event-based policy, omit this parameter.

$sel:eventSource:PolicyDetails', policyDetails_eventSource - The event that triggers the event-based policy.

This parameter is required for event-based policies only. If you are creating a snapshot or AMI policy, omit this parameter.

$sel:resourceTypes:PolicyDetails', policyDetails_resourceTypes - The target resource type for snapshot and AMI lifecycle policies. Use VOLUME to create snapshots of individual volumes or use INSTANCE to create multi-volume snapshots from the volumes for an instance.

This parameter is required for snapshot and AMI policies only. If you are creating an event-based policy, omit this parameter.

policyDetails_actions :: Lens' PolicyDetails (Maybe (NonEmpty Action))

The actions to be performed when the event-based policy is triggered. You can specify only one action per policy.

This parameter is required for event-based policies only. If you are creating a snapshot or AMI policy, omit this parameter.

policyDetails_targetTags :: Lens' PolicyDetails (Maybe (NonEmpty Tag))

The single tag that identifies targeted resources for this policy.

This parameter is required for snapshot and AMI policies only. If you are creating an event-based policy, omit this parameter.

policyDetails_policyType :: Lens' PolicyDetails (Maybe PolicyTypeValues)

The valid target resource types and actions a policy can manage. Specify EBS_SNAPSHOT_MANAGEMENT to create a lifecycle policy that manages the lifecycle of Amazon EBS snapshots. Specify IMAGE_MANAGEMENT to create a lifecycle policy that manages the lifecycle of EBS-backed AMIs. Specify EVENT_BASED_POLICY to create an event-based policy that performs specific actions when a defined event occurs in your Amazon Web Services account.

The default is EBS_SNAPSHOT_MANAGEMENT.

policyDetails_resourceLocations :: Lens' PolicyDetails (Maybe (NonEmpty ResourceLocationValues))

The location of the resources to backup. If the source resources are located in an Amazon Web Services Region, specify CLOUD. If the source resources are located on an Outpost in your account, specify OUTPOST.

If you specify OUTPOST, Amazon Data Lifecycle Manager backs up all resources of the specified type with matching target tags across all of the Outposts in your account.

policyDetails_parameters :: Lens' PolicyDetails (Maybe Parameters)

A set of optional parameters for snapshot and AMI lifecycle policies.

This parameter is required for snapshot and AMI policies only. If you are creating an event-based policy, omit this parameter.

policyDetails_schedules :: Lens' PolicyDetails (Maybe (NonEmpty Schedule))

The schedules of policy-defined actions for snapshot and AMI lifecycle policies. A policy can have up to four schedules—one mandatory schedule and up to three optional schedules.

This parameter is required for snapshot and AMI policies only. If you are creating an event-based policy, omit this parameter.

policyDetails_eventSource :: Lens' PolicyDetails (Maybe EventSource)

The event that triggers the event-based policy.

This parameter is required for event-based policies only. If you are creating a snapshot or AMI policy, omit this parameter.

policyDetails_resourceTypes :: Lens' PolicyDetails (Maybe (NonEmpty ResourceTypeValues))

The target resource type for snapshot and AMI lifecycle policies. Use VOLUME to create snapshots of individual volumes or use INSTANCE to create multi-volume snapshots from the volumes for an instance.

This parameter is required for snapshot and AMI policies only. If you are creating an event-based policy, omit this parameter.

RetainRule

data RetainRule

Specifies the retention rule for a lifecycle policy. You can retain snapshots based on either a count or a time interval.

See: newRetainRule smart constructor.

Constructors

RetainRule'

Fields

count :: Maybe Natural

The number of snapshots to retain for each volume, up to a maximum of 1000.

interval :: Maybe Natural

The amount of time to retain each snapshot. The maximum is 100 years. This is equivalent to 1200 months, 5200 weeks, or 36500 days.

intervalUnit :: Maybe RetentionIntervalUnitValues

The unit of time for time-based retention.

newRetainRule :: RetainRule

Create a value of RetainRule with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:count:RetainRule', retainRule_count - The number of snapshots to retain for each volume, up to a maximum of 1000.

$sel:interval:RetainRule', retainRule_interval - The amount of time to retain each snapshot. The maximum is 100 years. This is equivalent to 1200 months, 5200 weeks, or 36500 days.

$sel:intervalUnit:RetainRule', retainRule_intervalUnit - The unit of time for time-based retention.

retainRule_count :: Lens' RetainRule (Maybe Natural)

The number of snapshots to retain for each volume, up to a maximum of 1000.

retainRule_interval :: Lens' RetainRule (Maybe Natural)

The amount of time to retain each snapshot. The maximum is 100 years. This is equivalent to 1200 months, 5200 weeks, or 36500 days.

retainRule_intervalUnit :: Lens' RetainRule (Maybe RetentionIntervalUnitValues)

The unit of time for time-based retention.

Schedule

data Schedule

Specifies a backup schedule for a snapshot or AMI lifecycle policy.

See: newSchedule smart constructor.

Constructors

Schedule'

Fields

variableTags :: Maybe [Tag]

A collection of key/value pairs with values determined dynamically when the policy is executed. Keys may be any valid Amazon EC2 tag key. Values must be in one of the two following formats: $(instance-id) or $(timestamp). Variable tags are only valid for EBS Snapshot Management – Instance policies.

createRule :: Maybe CreateRule

The creation rule.

deprecateRule :: Maybe DeprecateRule

The AMI deprecation rule for the schedule.

copyTags :: Maybe Bool

Copy all user-defined tags on a source volume to snapshots of the volume created by this policy.

name :: Maybe Text

The name of the schedule.

shareRules :: Maybe [ShareRule]

The rule for sharing snapshots with other Amazon Web Services accounts.

tagsToAdd :: Maybe [Tag]

The tags to apply to policy-created resources. These user-defined tags are in addition to the Amazon Web Services-added lifecycle tags.

retainRule :: Maybe RetainRule

The retention rule.

crossRegionCopyRules :: Maybe [CrossRegionCopyRule]

The rule for cross-Region snapshot copies.

You can only specify cross-Region copy rules for policies that create snapshots in a Region. If the policy creates snapshots on an Outpost, then you cannot copy the snapshots to a Region or to an Outpost. If the policy creates snapshots in a Region, then snapshots can be copied to up to three Regions or Outposts.

fastRestoreRule :: Maybe FastRestoreRule

The rule for enabling fast snapshot restore.

newSchedule :: Schedule

Create a value of Schedule with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:variableTags:Schedule', schedule_variableTags - A collection of key/value pairs with values determined dynamically when the policy is executed. Keys may be any valid Amazon EC2 tag key. Values must be in one of the two following formats: $(instance-id) or $(timestamp). Variable tags are only valid for EBS Snapshot Management – Instance policies.

$sel:createRule:Schedule', schedule_createRule - The creation rule.

$sel:deprecateRule:Schedule', schedule_deprecateRule - The AMI deprecation rule for the schedule.

$sel:copyTags:Schedule', schedule_copyTags - Copy all user-defined tags on a source volume to snapshots of the volume created by this policy.

$sel:name:Schedule', schedule_name - The name of the schedule.

$sel:shareRules:Schedule', schedule_shareRules - The rule for sharing snapshots with other Amazon Web Services accounts.

$sel:tagsToAdd:Schedule', schedule_tagsToAdd - The tags to apply to policy-created resources. These user-defined tags are in addition to the Amazon Web Services-added lifecycle tags.

$sel:retainRule:Schedule', schedule_retainRule - The retention rule.

$sel:crossRegionCopyRules:Schedule', schedule_crossRegionCopyRules - The rule for cross-Region snapshot copies.

You can only specify cross-Region copy rules for policies that create snapshots in a Region. If the policy creates snapshots on an Outpost, then you cannot copy the snapshots to a Region or to an Outpost. If the policy creates snapshots in a Region, then snapshots can be copied to up to three Regions or Outposts.

$sel:fastRestoreRule:Schedule', schedule_fastRestoreRule - The rule for enabling fast snapshot restore.

schedule_variableTags :: Lens' Schedule (Maybe [Tag])

A collection of key/value pairs with values determined dynamically when the policy is executed. Keys may be any valid Amazon EC2 tag key. Values must be in one of the two following formats: $(instance-id) or $(timestamp). Variable tags are only valid for EBS Snapshot Management – Instance policies.

schedule_createRule :: Lens' Schedule (Maybe CreateRule)

The creation rule.

schedule_deprecateRule :: Lens' Schedule (Maybe DeprecateRule)

The AMI deprecation rule for the schedule.

schedule_copyTags :: Lens' Schedule (Maybe Bool)

Copy all user-defined tags on a source volume to snapshots of the volume created by this policy.

schedule_name :: Lens' Schedule (Maybe Text)

The name of the schedule.

schedule_shareRules :: Lens' Schedule (Maybe [ShareRule])

The rule for sharing snapshots with other Amazon Web Services accounts.

schedule_tagsToAdd :: Lens' Schedule (Maybe [Tag])

The tags to apply to policy-created resources. These user-defined tags are in addition to the Amazon Web Services-added lifecycle tags.

schedule_retainRule :: Lens' Schedule (Maybe RetainRule)

The retention rule.

schedule_crossRegionCopyRules :: Lens' Schedule (Maybe [CrossRegionCopyRule])

The rule for cross-Region snapshot copies.

You can only specify cross-Region copy rules for policies that create snapshots in a Region. If the policy creates snapshots on an Outpost, then you cannot copy the snapshots to a Region or to an Outpost. If the policy creates snapshots in a Region, then snapshots can be copied to up to three Regions or Outposts.

schedule_fastRestoreRule :: Lens' Schedule (Maybe FastRestoreRule)

The rule for enabling fast snapshot restore.

ShareRule

data ShareRule

Specifies a rule for sharing snapshots across Amazon Web Services accounts.

See: newShareRule smart constructor.

Constructors

ShareRule'

Fields

unshareIntervalUnit :: Maybe RetentionIntervalUnitValues

The unit of time for the automatic unsharing interval.

unshareInterval :: Maybe Natural

The period after which snapshots that are shared with other Amazon Web Services accounts are automatically unshared.

targetAccounts :: NonEmpty Text

The IDs of the Amazon Web Services accounts with which to share the snapshots.

newShareRule

Arguments

:: NonEmpty Text
$sel:targetAccounts:ShareRule'

-> ShareRule

Create a value of ShareRule with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:unshareIntervalUnit:ShareRule', shareRule_unshareIntervalUnit - The unit of time for the automatic unsharing interval.

$sel:unshareInterval:ShareRule', shareRule_unshareInterval - The period after which snapshots that are shared with other Amazon Web Services accounts are automatically unshared.

$sel:targetAccounts:ShareRule', shareRule_targetAccounts - The IDs of the Amazon Web Services accounts with which to share the snapshots.

shareRule_unshareIntervalUnit :: Lens' ShareRule (Maybe RetentionIntervalUnitValues)

The unit of time for the automatic unsharing interval.

shareRule_unshareInterval :: Lens' ShareRule (Maybe Natural)

The period after which snapshots that are shared with other Amazon Web Services accounts are automatically unshared.

shareRule_targetAccounts :: Lens' ShareRule (NonEmpty Text)

The IDs of the Amazon Web Services accounts with which to share the snapshots.

Tag

data Tag

Specifies a tag for a resource.

See: newTag smart constructor.

Constructors

Tag'

Fields

key :: Text

The tag key.

value :: Text

The tag value.

newTag

Arguments

:: Text
$sel:key:Tag'

-> Text
$sel:value:Tag'

-> Tag

Create a value of Tag with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:key:Tag', tag_key - The tag key.

$sel:value:Tag', tag_value - The tag value.

tag_key :: Lens' Tag Text

The tag key.

tag_value :: Lens' Tag Text

The tag value.