Amazonka.DLM.Types.CreateRule

Description

Documentation

data CreateRule

Specifies when to create snapshots of EBS volumes.

You must specify either a Cron expression or an interval, interval unit, and start time. You cannot specify both.

See: newCreateRule smart constructor.

Constructors

CreateRule'

Fields

location :: Maybe LocationValues

Specifies the destination for snapshots created by the policy. To create snapshots in the same Region as the source resource, specify CLOUD. To create snapshots on the same Outpost as the source resource, specify OUTPOST_LOCAL. If you omit this parameter, CLOUD is used by default.

If the policy targets resources in an Amazon Web Services Region, then you must create snapshots in the same Region as the source resource.

If the policy targets resources on an Outpost, then you can create snapshots on the same Outpost as the source resource, or in the Region of that Outpost.

interval :: Maybe Natural

The interval between snapshots. The supported values are 1, 2, 3, 4, 6, 8, 12, and 24.

cronExpression :: Maybe Text

The schedule, as a Cron expression. The schedule interval must be between 1 hour and 1 year. For more information, see Cron expressions in the Amazon CloudWatch User Guide.

times :: Maybe [Text]

The time, in UTC, to start the operation. The supported format is hh:mm.

The operation occurs within a one-hour window following the specified time. If you do not specify a time, Amazon DLM selects a time within the next 24 hours.

intervalUnit :: Maybe IntervalUnitValues

The interval unit.

newCreateRule :: CreateRule

Create a value of CreateRule with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:location:CreateRule', createRule_location - Specifies the destination for snapshots created by the policy. To create snapshots in the same Region as the source resource, specify CLOUD. To create snapshots on the same Outpost as the source resource, specify OUTPOST_LOCAL. If you omit this parameter, CLOUD is used by default.

If the policy targets resources in an Amazon Web Services Region, then you must create snapshots in the same Region as the source resource.

If the policy targets resources on an Outpost, then you can create snapshots on the same Outpost as the source resource, or in the Region of that Outpost.

$sel:interval:CreateRule', createRule_interval - The interval between snapshots. The supported values are 1, 2, 3, 4, 6, 8, 12, and 24.

$sel:cronExpression:CreateRule', createRule_cronExpression - The schedule, as a Cron expression. The schedule interval must be between 1 hour and 1 year. For more information, see Cron expressions in the Amazon CloudWatch User Guide.

$sel:times:CreateRule', createRule_times - The time, in UTC, to start the operation. The supported format is hh:mm.

The operation occurs within a one-hour window following the specified time. If you do not specify a time, Amazon DLM selects a time within the next 24 hours.

$sel:intervalUnit:CreateRule', createRule_intervalUnit - The interval unit.

createRule_location :: Lens' CreateRule (Maybe LocationValues)

Specifies the destination for snapshots created by the policy. To create snapshots in the same Region as the source resource, specify CLOUD. To create snapshots on the same Outpost as the source resource, specify OUTPOST_LOCAL. If you omit this parameter, CLOUD is used by default.

If the policy targets resources in an Amazon Web Services Region, then you must create snapshots in the same Region as the source resource.

If the policy targets resources on an Outpost, then you can create snapshots on the same Outpost as the source resource, or in the Region of that Outpost.

createRule_interval :: Lens' CreateRule (Maybe Natural)

The interval between snapshots. The supported values are 1, 2, 3, 4, 6, 8, 12, and 24.

createRule_cronExpression :: Lens' CreateRule (Maybe Text)

The schedule, as a Cron expression. The schedule interval must be between 1 hour and 1 year. For more information, see Cron expressions in the Amazon CloudWatch User Guide.

createRule_times :: Lens' CreateRule (Maybe [Text])

The time, in UTC, to start the operation. The supported format is hh:mm.

The operation occurs within a one-hour window following the specified time. If you do not specify a time, Amazon DLM selects a time within the next 24 hours.

createRule_intervalUnit :: Lens' CreateRule (Maybe IntervalUnitValues)

The interval unit.