Amazonka.DLM.Types.CrossRegionCopyRule

Description

Documentation

data CrossRegionCopyRule

Specifies a rule for cross-Region snapshot copies.

See: newCrossRegionCopyRule smart constructor.

Constructors

CrossRegionCopyRule'

Fields

deprecateRule :: Maybe CrossRegionCopyDeprecateRule

The AMI deprecation rule for cross-Region AMI copies created by the rule.

targetRegion :: Maybe Text

Avoid using this parameter when creating new policies. Instead, use Target to specify a target Region or a target Outpost for snapshot copies.

For policies created before the Target parameter was introduced, this parameter indicates the target Region for snapshot copies.

copyTags :: Maybe Bool

Indicates whether to copy all user-defined tags from the source snapshot to the cross-Region snapshot copy.

cmkArn :: Maybe Text

The Amazon Resource Name (ARN) of the KMS key to use for EBS encryption. If this parameter is not specified, the default KMS key for the account is used.

retainRule :: Maybe CrossRegionCopyRetainRule

The retention rule that indicates how long snapshot copies are to be retained in the destination Region.

target :: Maybe Text

The target Region or the Amazon Resource Name (ARN) of the target Outpost for the snapshot copies.

Use this parameter instead of TargetRegion. Do not specify both.

encrypted :: Bool

To encrypt a copy of an unencrypted snapshot if encryption by default is not enabled, enable encryption using this parameter. Copies of encrypted snapshots are encrypted, even if this parameter is false or if encryption by default is not enabled.

newCrossRegionCopyRule

Arguments

:: Bool
$sel:encrypted:CrossRegionCopyRule'

-> CrossRegionCopyRule

Create a value of CrossRegionCopyRule with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:deprecateRule:CrossRegionCopyRule', crossRegionCopyRule_deprecateRule - The AMI deprecation rule for cross-Region AMI copies created by the rule.

$sel:targetRegion:CrossRegionCopyRule', crossRegionCopyRule_targetRegion - Avoid using this parameter when creating new policies. Instead, use Target to specify a target Region or a target Outpost for snapshot copies.

For policies created before the Target parameter was introduced, this parameter indicates the target Region for snapshot copies.

$sel:copyTags:CrossRegionCopyRule', crossRegionCopyRule_copyTags - Indicates whether to copy all user-defined tags from the source snapshot to the cross-Region snapshot copy.

$sel:cmkArn:CrossRegionCopyRule', crossRegionCopyRule_cmkArn - The Amazon Resource Name (ARN) of the KMS key to use for EBS encryption. If this parameter is not specified, the default KMS key for the account is used.

$sel:retainRule:CrossRegionCopyRule', crossRegionCopyRule_retainRule - The retention rule that indicates how long snapshot copies are to be retained in the destination Region.

$sel:target:CrossRegionCopyRule', crossRegionCopyRule_target - The target Region or the Amazon Resource Name (ARN) of the target Outpost for the snapshot copies.

Use this parameter instead of TargetRegion. Do not specify both.

$sel:encrypted:CrossRegionCopyRule', crossRegionCopyRule_encrypted - To encrypt a copy of an unencrypted snapshot if encryption by default is not enabled, enable encryption using this parameter. Copies of encrypted snapshots are encrypted, even if this parameter is false or if encryption by default is not enabled.

crossRegionCopyRule_deprecateRule :: Lens' CrossRegionCopyRule (Maybe CrossRegionCopyDeprecateRule)

The AMI deprecation rule for cross-Region AMI copies created by the rule.

crossRegionCopyRule_targetRegion :: Lens' CrossRegionCopyRule (Maybe Text)

Avoid using this parameter when creating new policies. Instead, use Target to specify a target Region or a target Outpost for snapshot copies.

For policies created before the Target parameter was introduced, this parameter indicates the target Region for snapshot copies.

crossRegionCopyRule_copyTags :: Lens' CrossRegionCopyRule (Maybe Bool)

Indicates whether to copy all user-defined tags from the source snapshot to the cross-Region snapshot copy.

crossRegionCopyRule_cmkArn :: Lens' CrossRegionCopyRule (Maybe Text)

The Amazon Resource Name (ARN) of the KMS key to use for EBS encryption. If this parameter is not specified, the default KMS key for the account is used.

crossRegionCopyRule_retainRule :: Lens' CrossRegionCopyRule (Maybe CrossRegionCopyRetainRule)

The retention rule that indicates how long snapshot copies are to be retained in the destination Region.

crossRegionCopyRule_target :: Lens' CrossRegionCopyRule (Maybe Text)

The target Region or the Amazon Resource Name (ARN) of the target Outpost for the snapshot copies.

Use this parameter instead of TargetRegion. Do not specify both.

crossRegionCopyRule_encrypted :: Lens' CrossRegionCopyRule Bool

To encrypt a copy of an unencrypted snapshot if encryption by default is not enabled, enable encryption using this parameter. Copies of encrypted snapshots are encrypted, even if this parameter is false or if encryption by default is not enabled.