Amazonka.DLM.Types.FastRestoreRule

Description

Documentation

data FastRestoreRule

Specifies a rule for enabling fast snapshot restore. You can enable fast snapshot restore based on either a count or a time interval.

See: newFastRestoreRule smart constructor.

Constructors

FastRestoreRule'

Fields

count :: Maybe Natural

The number of snapshots to be enabled with fast snapshot restore.

interval :: Maybe Natural

The amount of time to enable fast snapshot restore. The maximum is 100 years. This is equivalent to 1200 months, 5200 weeks, or 36500 days.

intervalUnit :: Maybe RetentionIntervalUnitValues

The unit of time for enabling fast snapshot restore.

availabilityZones :: NonEmpty Text

The Availability Zones in which to enable fast snapshot restore.

newFastRestoreRule

Arguments

:: NonEmpty Text
$sel:availabilityZones:FastRestoreRule'

-> FastRestoreRule

Create a value of FastRestoreRule with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:count:FastRestoreRule', fastRestoreRule_count - The number of snapshots to be enabled with fast snapshot restore.

$sel:interval:FastRestoreRule', fastRestoreRule_interval - The amount of time to enable fast snapshot restore. The maximum is 100 years. This is equivalent to 1200 months, 5200 weeks, or 36500 days.

$sel:intervalUnit:FastRestoreRule', fastRestoreRule_intervalUnit - The unit of time for enabling fast snapshot restore.

$sel:availabilityZones:FastRestoreRule', fastRestoreRule_availabilityZones - The Availability Zones in which to enable fast snapshot restore.

fastRestoreRule_count :: Lens' FastRestoreRule (Maybe Natural)

The number of snapshots to be enabled with fast snapshot restore.

fastRestoreRule_interval :: Lens' FastRestoreRule (Maybe Natural)

The amount of time to enable fast snapshot restore. The maximum is 100 years. This is equivalent to 1200 months, 5200 weeks, or 36500 days.

fastRestoreRule_intervalUnit :: Lens' FastRestoreRule (Maybe RetentionIntervalUnitValues)

The unit of time for enabling fast snapshot restore.

fastRestoreRule_availabilityZones :: Lens' FastRestoreRule (NonEmpty Text)

The Availability Zones in which to enable fast snapshot restore.