Amazonka.DLM.Types.LifecyclePolicy

Description

Documentation

data LifecyclePolicy

Detailed information about a lifecycle policy.

See: newLifecyclePolicy smart constructor.

Constructors

LifecyclePolicy'

Fields

state :: Maybe GettablePolicyStateValues

The activation state of the lifecycle policy.

policyDetails :: Maybe PolicyDetails

The configuration of the lifecycle policy

policyId :: Maybe Text

The identifier of the lifecycle policy.

executionRoleArn :: Maybe Text

The Amazon Resource Name (ARN) of the IAM role used to run the operations specified by the lifecycle policy.

dateCreated :: Maybe POSIX

The local date and time when the lifecycle policy was created.

statusMessage :: Maybe Text

The description of the status.

dateModified :: Maybe POSIX

The local date and time when the lifecycle policy was last modified.

policyArn :: Maybe Text

The Amazon Resource Name (ARN) of the policy.

description :: Maybe Text

The description of the lifecycle policy.

tags :: Maybe (HashMap Text Text)

The tags.

newLifecyclePolicy :: LifecyclePolicy

Create a value of LifecyclePolicy with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:state:LifecyclePolicy', lifecyclePolicy_state - The activation state of the lifecycle policy.

$sel:policyDetails:LifecyclePolicy', lifecyclePolicy_policyDetails - The configuration of the lifecycle policy

$sel:policyId:LifecyclePolicy', lifecyclePolicy_policyId - The identifier of the lifecycle policy.

$sel:executionRoleArn:LifecyclePolicy', lifecyclePolicy_executionRoleArn - The Amazon Resource Name (ARN) of the IAM role used to run the operations specified by the lifecycle policy.

$sel:dateCreated:LifecyclePolicy', lifecyclePolicy_dateCreated - The local date and time when the lifecycle policy was created.

$sel:statusMessage:LifecyclePolicy', lifecyclePolicy_statusMessage - The description of the status.

$sel:dateModified:LifecyclePolicy', lifecyclePolicy_dateModified - The local date and time when the lifecycle policy was last modified.

$sel:policyArn:LifecyclePolicy', lifecyclePolicy_policyArn - The Amazon Resource Name (ARN) of the policy.

$sel:description:LifecyclePolicy', lifecyclePolicy_description - The description of the lifecycle policy.

$sel:tags:LifecyclePolicy', lifecyclePolicy_tags - The tags.

lifecyclePolicy_state :: Lens' LifecyclePolicy (Maybe GettablePolicyStateValues)

The activation state of the lifecycle policy.

lifecyclePolicy_policyDetails :: Lens' LifecyclePolicy (Maybe PolicyDetails)

The configuration of the lifecycle policy

lifecyclePolicy_policyId :: Lens' LifecyclePolicy (Maybe Text)

The identifier of the lifecycle policy.

lifecyclePolicy_executionRoleArn :: Lens' LifecyclePolicy (Maybe Text)

The Amazon Resource Name (ARN) of the IAM role used to run the operations specified by the lifecycle policy.

lifecyclePolicy_dateCreated :: Lens' LifecyclePolicy (Maybe UTCTime)

The local date and time when the lifecycle policy was created.

lifecyclePolicy_statusMessage :: Lens' LifecyclePolicy (Maybe Text)

The description of the status.

lifecyclePolicy_dateModified :: Lens' LifecyclePolicy (Maybe UTCTime)

The local date and time when the lifecycle policy was last modified.

lifecyclePolicy_policyArn :: Lens' LifecyclePolicy (Maybe Text)

The Amazon Resource Name (ARN) of the policy.

lifecyclePolicy_description :: Lens' LifecyclePolicy (Maybe Text)

The description of the lifecycle policy.

lifecyclePolicy_tags :: Lens' LifecyclePolicy (Maybe (HashMap Text Text))

The tags.