Amazonka.DLM.Types.Parameters

Description

Documentation

data Parameters

Specifies optional parameters to add to a policy. The set of valid parameters depends on the combination of policy type and resource type.

See: newParameters smart constructor.

Constructors

Parameters'

Fields

noReboot :: Maybe Bool

Applies to AMI lifecycle policies only. Indicates whether targeted instances are rebooted when the lifecycle policy runs. true indicates that targeted instances are not rebooted when the policy runs. false indicates that target instances are rebooted when the policy runs. The default is true (instances are not rebooted).

excludeBootVolume :: Maybe Bool

EBS Snapshot Management – Instance policies only

Indicates whether to exclude the root volume from snapshots created using CreateSnapshots. The default is false.

newParameters :: Parameters

Create a value of Parameters with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:noReboot:Parameters', parameters_noReboot - Applies to AMI lifecycle policies only. Indicates whether targeted instances are rebooted when the lifecycle policy runs. true indicates that targeted instances are not rebooted when the policy runs. false indicates that target instances are rebooted when the policy runs. The default is true (instances are not rebooted).

$sel:excludeBootVolume:Parameters', parameters_excludeBootVolume - [EBS Snapshot Management – Instance policies only] Indicates whether to exclude the root volume from snapshots created using CreateSnapshots. The default is false.

parameters_noReboot :: Lens' Parameters (Maybe Bool)

Applies to AMI lifecycle policies only. Indicates whether targeted instances are rebooted when the lifecycle policy runs. true indicates that targeted instances are not rebooted when the policy runs. false indicates that target instances are rebooted when the policy runs. The default is true (instances are not rebooted).

parameters_excludeBootVolume :: Lens' Parameters (Maybe Bool)

EBS Snapshot Management – Instance policies only

Indicates whether to exclude the root volume from snapshots created using CreateSnapshots. The default is false.