Amazonka.DLM.Types.Schedule

Description

Documentation

data Schedule

Specifies a backup schedule for a snapshot or AMI lifecycle policy.

See: newSchedule smart constructor.

Constructors

Schedule'

Fields

variableTags :: Maybe [Tag]

A collection of key/value pairs with values determined dynamically when the policy is executed. Keys may be any valid Amazon EC2 tag key. Values must be in one of the two following formats: $(instance-id) or $(timestamp). Variable tags are only valid for EBS Snapshot Management – Instance policies.

createRule :: Maybe CreateRule

The creation rule.

deprecateRule :: Maybe DeprecateRule

The AMI deprecation rule for the schedule.

copyTags :: Maybe Bool

Copy all user-defined tags on a source volume to snapshots of the volume created by this policy.

name :: Maybe Text

The name of the schedule.

shareRules :: Maybe [ShareRule]

The rule for sharing snapshots with other Amazon Web Services accounts.

tagsToAdd :: Maybe [Tag]

The tags to apply to policy-created resources. These user-defined tags are in addition to the Amazon Web Services-added lifecycle tags.

retainRule :: Maybe RetainRule

The retention rule.

crossRegionCopyRules :: Maybe [CrossRegionCopyRule]

The rule for cross-Region snapshot copies.

You can only specify cross-Region copy rules for policies that create snapshots in a Region. If the policy creates snapshots on an Outpost, then you cannot copy the snapshots to a Region or to an Outpost. If the policy creates snapshots in a Region, then snapshots can be copied to up to three Regions or Outposts.

fastRestoreRule :: Maybe FastRestoreRule

The rule for enabling fast snapshot restore.

newSchedule :: Schedule

Create a value of Schedule with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:variableTags:Schedule', schedule_variableTags - A collection of key/value pairs with values determined dynamically when the policy is executed. Keys may be any valid Amazon EC2 tag key. Values must be in one of the two following formats: $(instance-id) or $(timestamp). Variable tags are only valid for EBS Snapshot Management – Instance policies.

$sel:createRule:Schedule', schedule_createRule - The creation rule.

$sel:deprecateRule:Schedule', schedule_deprecateRule - The AMI deprecation rule for the schedule.

$sel:copyTags:Schedule', schedule_copyTags - Copy all user-defined tags on a source volume to snapshots of the volume created by this policy.

$sel:name:Schedule', schedule_name - The name of the schedule.

$sel:shareRules:Schedule', schedule_shareRules - The rule for sharing snapshots with other Amazon Web Services accounts.

$sel:tagsToAdd:Schedule', schedule_tagsToAdd - The tags to apply to policy-created resources. These user-defined tags are in addition to the Amazon Web Services-added lifecycle tags.

$sel:retainRule:Schedule', schedule_retainRule - The retention rule.

$sel:crossRegionCopyRules:Schedule', schedule_crossRegionCopyRules - The rule for cross-Region snapshot copies.

You can only specify cross-Region copy rules for policies that create snapshots in a Region. If the policy creates snapshots on an Outpost, then you cannot copy the snapshots to a Region or to an Outpost. If the policy creates snapshots in a Region, then snapshots can be copied to up to three Regions or Outposts.

$sel:fastRestoreRule:Schedule', schedule_fastRestoreRule - The rule for enabling fast snapshot restore.

schedule_variableTags :: Lens' Schedule (Maybe [Tag])

A collection of key/value pairs with values determined dynamically when the policy is executed. Keys may be any valid Amazon EC2 tag key. Values must be in one of the two following formats: $(instance-id) or $(timestamp). Variable tags are only valid for EBS Snapshot Management – Instance policies.

schedule_createRule :: Lens' Schedule (Maybe CreateRule)

The creation rule.

schedule_deprecateRule :: Lens' Schedule (Maybe DeprecateRule)

The AMI deprecation rule for the schedule.

schedule_copyTags :: Lens' Schedule (Maybe Bool)

Copy all user-defined tags on a source volume to snapshots of the volume created by this policy.

schedule_name :: Lens' Schedule (Maybe Text)

The name of the schedule.

schedule_shareRules :: Lens' Schedule (Maybe [ShareRule])

The rule for sharing snapshots with other Amazon Web Services accounts.

schedule_tagsToAdd :: Lens' Schedule (Maybe [Tag])

The tags to apply to policy-created resources. These user-defined tags are in addition to the Amazon Web Services-added lifecycle tags.

schedule_retainRule :: Lens' Schedule (Maybe RetainRule)

The retention rule.

schedule_crossRegionCopyRules :: Lens' Schedule (Maybe [CrossRegionCopyRule])

The rule for cross-Region snapshot copies.

You can only specify cross-Region copy rules for policies that create snapshots in a Region. If the policy creates snapshots on an Outpost, then you cannot copy the snapshots to a Region or to an Outpost. If the policy creates snapshots in a Region, then snapshots can be copied to up to three Regions or Outposts.

schedule_fastRestoreRule :: Lens' Schedule (Maybe FastRestoreRule)

The rule for enabling fast snapshot restore.