Amazonka.DLM.UpdateLifecyclePolicy

Description

Updates the specified lifecycle policy.

Creating a Request

data UpdateLifecyclePolicy

See: newUpdateLifecyclePolicy smart constructor.

Constructors

UpdateLifecyclePolicy'

Fields

state :: Maybe SettablePolicyStateValues

The desired activation state of the lifecycle policy after creation.

policyDetails :: Maybe PolicyDetails

The configuration of the lifecycle policy. You cannot update the policy type or the resource type.

executionRoleArn :: Maybe Text

The Amazon Resource Name (ARN) of the IAM role used to run the operations specified by the lifecycle policy.

description :: Maybe Text

A description of the lifecycle policy.

policyId :: Text

The identifier of the lifecycle policy.

newUpdateLifecyclePolicy

Arguments

:: Text
$sel:policyId:UpdateLifecyclePolicy'

-> UpdateLifecyclePolicy

Create a value of UpdateLifecyclePolicy with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:state:UpdateLifecyclePolicy', updateLifecyclePolicy_state - The desired activation state of the lifecycle policy after creation.

$sel:policyDetails:UpdateLifecyclePolicy', updateLifecyclePolicy_policyDetails - The configuration of the lifecycle policy. You cannot update the policy type or the resource type.

$sel:executionRoleArn:UpdateLifecyclePolicy', updateLifecyclePolicy_executionRoleArn - The Amazon Resource Name (ARN) of the IAM role used to run the operations specified by the lifecycle policy.

$sel:description:UpdateLifecyclePolicy', updateLifecyclePolicy_description - A description of the lifecycle policy.

$sel:policyId:UpdateLifecyclePolicy', updateLifecyclePolicy_policyId - The identifier of the lifecycle policy.

Request Lenses

updateLifecyclePolicy_state :: Lens' UpdateLifecyclePolicy (Maybe SettablePolicyStateValues)

The desired activation state of the lifecycle policy after creation.

updateLifecyclePolicy_policyDetails :: Lens' UpdateLifecyclePolicy (Maybe PolicyDetails)

The configuration of the lifecycle policy. You cannot update the policy type or the resource type.

updateLifecyclePolicy_executionRoleArn :: Lens' UpdateLifecyclePolicy (Maybe Text)

The Amazon Resource Name (ARN) of the IAM role used to run the operations specified by the lifecycle policy.

updateLifecyclePolicy_description :: Lens' UpdateLifecyclePolicy (Maybe Text)

A description of the lifecycle policy.

updateLifecyclePolicy_policyId :: Lens' UpdateLifecyclePolicy Text

The identifier of the lifecycle policy.

Destructuring the Response

data UpdateLifecyclePolicyResponse

See: newUpdateLifecyclePolicyResponse smart constructor.

Constructors

UpdateLifecyclePolicyResponse'

Fields

httpStatus :: Int

The response's http status code.

newUpdateLifecyclePolicyResponse

Arguments

:: Int
$sel:httpStatus:UpdateLifecyclePolicyResponse'

-> UpdateLifecyclePolicyResponse

Create a value of UpdateLifecyclePolicyResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:httpStatus:UpdateLifecyclePolicyResponse', updateLifecyclePolicyResponse_httpStatus - The response's http status code.

Response Lenses

updateLifecyclePolicyResponse_httpStatus :: Lens' UpdateLifecyclePolicyResponse Int

The response's http status code.