Amazonka.DMS.CreateReplicationTask

Description

Creates a replication task using the specified parameters.

Creating a Request

data CreateReplicationTask

See: newCreateReplicationTask smart constructor.

Constructors

CreateReplicationTask'

Fields

replicationTaskSettings :: Maybe Text

Overall settings for the task, in JSON format. For more information, see Specifying Task Settings for Database Migration Service Tasks in the Database Migration Service User Guide.

cdcStartPosition :: Maybe Text

Indicates when you want a change data capture (CDC) operation to start. Use either CdcStartPosition or CdcStartTime to specify when you want a CDC operation to start. Specifying both values results in an error.

The value can be in date, checkpoint, or LSN/SCN format.

Date Example: --cdc-start-position “2018-03-08T12:12:12”

Checkpoint Example: --cdc-start-position "checkpoint:V1mysql-bin-changelog.157832:1975:-1:2002:677883278264080:mysql-bin-changelog.157832:187600#93"

LSN Example: --cdc-start-position “mysql-bin-changelog.000024:373”

When you use this task setting with a source PostgreSQL database, a logical replication slot should already be created and associated with the source endpoint. You can verify this by setting the slotName extra connection attribute to the name of this logical replication slot. For more information, see Extra Connection Attributes When Using PostgreSQL as a Source for DMS.

taskData :: Maybe Text

Supplemental information that the task requires to migrate the data for certain source and target endpoints. For more information, see Specifying Supplemental Data for Task Settings in the Database Migration Service User Guide.

cdcStopPosition :: Maybe Text

Indicates when you want a change data capture (CDC) operation to stop. The value can be either server time or commit time.

Server time example: --cdc-stop-position “server_time:2018-02-09T12:12:12”

Commit time example: --cdc-stop-position “commit_time: 2018-02-09T12:12:12 “

resourceIdentifier :: Maybe Text

A friendly name for the resource identifier at the end of the EndpointArn response parameter that is returned in the created Endpoint object. The value for this parameter can have up to 31 characters. It can contain only ASCII letters, digits, and hyphen ('-'). Also, it can't end with a hyphen or contain two consecutive hyphens, and can only begin with a letter, such as Example-App-ARN1. For example, this value might result in the EndpointArn value arn:aws:dms:eu-west-1:012345678901:rep:Example-App-ARN1. If you don't specify a ResourceIdentifier value, DMS generates a default identifier value for the end of EndpointArn.

tags :: Maybe [Tag]

One or more tags to be assigned to the replication task.

cdcStartTime :: Maybe POSIX

Indicates the start time for a change data capture (CDC) operation. Use either CdcStartTime or CdcStartPosition to specify when you want a CDC operation to start. Specifying both values results in an error.

Timestamp Example: --cdc-start-time “2018-03-08T12:12:12”

replicationTaskIdentifier :: Text

An identifier for the replication task.

Constraints:

Must contain 1-255 alphanumeric characters or hyphens.

First character must be a letter.

Cannot end with a hyphen or contain two consecutive hyphens.

sourceEndpointArn :: Text

An Amazon Resource Name (ARN) that uniquely identifies the source endpoint.

targetEndpointArn :: Text

An Amazon Resource Name (ARN) that uniquely identifies the target endpoint.

replicationInstanceArn :: Text

The Amazon Resource Name (ARN) of a replication instance.

migrationType :: MigrationTypeValue

The migration type. Valid values: full-load | cdc | full-load-and-cdc

tableMappings :: Text

The table mappings for the task, in JSON format. For more information, see Using Table Mapping to Specify Task Settings in the Database Migration Service User Guide.

newCreateReplicationTask

Arguments

:: Text
$sel:replicationTaskIdentifier:CreateReplicationTask'

-> Text
$sel:sourceEndpointArn:CreateReplicationTask'

-> Text
$sel:targetEndpointArn:CreateReplicationTask'

-> Text
$sel:replicationInstanceArn:CreateReplicationTask'

-> MigrationTypeValue
$sel:migrationType:CreateReplicationTask'

-> Text
$sel:tableMappings:CreateReplicationTask'

-> CreateReplicationTask

Create a value of CreateReplicationTask with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:replicationTaskSettings:CreateReplicationTask', createReplicationTask_replicationTaskSettings - Overall settings for the task, in JSON format. For more information, see Specifying Task Settings for Database Migration Service Tasks in the Database Migration Service User Guide.

$sel:cdcStartPosition:CreateReplicationTask', createReplicationTask_cdcStartPosition - Indicates when you want a change data capture (CDC) operation to start. Use either CdcStartPosition or CdcStartTime to specify when you want a CDC operation to start. Specifying both values results in an error.

The value can be in date, checkpoint, or LSN/SCN format.

Date Example: --cdc-start-position “2018-03-08T12:12:12”

Checkpoint Example: --cdc-start-position "checkpoint:V1mysql-bin-changelog.157832:1975:-1:2002:677883278264080:mysql-bin-changelog.157832:187600#93"

LSN Example: --cdc-start-position “mysql-bin-changelog.000024:373”

When you use this task setting with a source PostgreSQL database, a logical replication slot should already be created and associated with the source endpoint. You can verify this by setting the slotName extra connection attribute to the name of this logical replication slot. For more information, see Extra Connection Attributes When Using PostgreSQL as a Source for DMS.

$sel:taskData:CreateReplicationTask', createReplicationTask_taskData - Supplemental information that the task requires to migrate the data for certain source and target endpoints. For more information, see Specifying Supplemental Data for Task Settings in the Database Migration Service User Guide.

$sel:cdcStopPosition:CreateReplicationTask', createReplicationTask_cdcStopPosition - Indicates when you want a change data capture (CDC) operation to stop. The value can be either server time or commit time.

Server time example: --cdc-stop-position “server_time:2018-02-09T12:12:12”

Commit time example: --cdc-stop-position “commit_time: 2018-02-09T12:12:12 “

$sel:resourceIdentifier:CreateReplicationTask', createReplicationTask_resourceIdentifier - A friendly name for the resource identifier at the end of the EndpointArn response parameter that is returned in the created Endpoint object. The value for this parameter can have up to 31 characters. It can contain only ASCII letters, digits, and hyphen ('-'). Also, it can't end with a hyphen or contain two consecutive hyphens, and can only begin with a letter, such as Example-App-ARN1. For example, this value might result in the EndpointArn value arn:aws:dms:eu-west-1:012345678901:rep:Example-App-ARN1. If you don't specify a ResourceIdentifier value, DMS generates a default identifier value for the end of EndpointArn.

$sel:tags:CreateReplicationTask', createReplicationTask_tags - One or more tags to be assigned to the replication task.

$sel:cdcStartTime:CreateReplicationTask', createReplicationTask_cdcStartTime - Indicates the start time for a change data capture (CDC) operation. Use either CdcStartTime or CdcStartPosition to specify when you want a CDC operation to start. Specifying both values results in an error.

Timestamp Example: --cdc-start-time “2018-03-08T12:12:12”

$sel:replicationTaskIdentifier:CreateReplicationTask', createReplicationTask_replicationTaskIdentifier - An identifier for the replication task.

Constraints:

Must contain 1-255 alphanumeric characters or hyphens.

First character must be a letter.

Cannot end with a hyphen or contain two consecutive hyphens.

$sel:sourceEndpointArn:CreateReplicationTask', createReplicationTask_sourceEndpointArn - An Amazon Resource Name (ARN) that uniquely identifies the source endpoint.

$sel:targetEndpointArn:CreateReplicationTask', createReplicationTask_targetEndpointArn - An Amazon Resource Name (ARN) that uniquely identifies the target endpoint.

$sel:replicationInstanceArn:CreateReplicationTask', createReplicationTask_replicationInstanceArn - The Amazon Resource Name (ARN) of a replication instance.

$sel:migrationType:CreateReplicationTask', createReplicationTask_migrationType - The migration type. Valid values: full-load | cdc | full-load-and-cdc

$sel:tableMappings:CreateReplicationTask', createReplicationTask_tableMappings - The table mappings for the task, in JSON format. For more information, see Using Table Mapping to Specify Task Settings in the Database Migration Service User Guide.

Request Lenses

createReplicationTask_replicationTaskSettings :: Lens' CreateReplicationTask (Maybe Text)

Overall settings for the task, in JSON format. For more information, see Specifying Task Settings for Database Migration Service Tasks in the Database Migration Service User Guide.

createReplicationTask_cdcStartPosition :: Lens' CreateReplicationTask (Maybe Text)

Indicates when you want a change data capture (CDC) operation to start. Use either CdcStartPosition or CdcStartTime to specify when you want a CDC operation to start. Specifying both values results in an error.

The value can be in date, checkpoint, or LSN/SCN format.

Date Example: --cdc-start-position “2018-03-08T12:12:12”

Checkpoint Example: --cdc-start-position "checkpoint:V1mysql-bin-changelog.157832:1975:-1:2002:677883278264080:mysql-bin-changelog.157832:187600#93"

LSN Example: --cdc-start-position “mysql-bin-changelog.000024:373”

When you use this task setting with a source PostgreSQL database, a logical replication slot should already be created and associated with the source endpoint. You can verify this by setting the slotName extra connection attribute to the name of this logical replication slot. For more information, see Extra Connection Attributes When Using PostgreSQL as a Source for DMS.

createReplicationTask_taskData :: Lens' CreateReplicationTask (Maybe Text)

Supplemental information that the task requires to migrate the data for certain source and target endpoints. For more information, see Specifying Supplemental Data for Task Settings in the Database Migration Service User Guide.

createReplicationTask_cdcStopPosition :: Lens' CreateReplicationTask (Maybe Text)

Indicates when you want a change data capture (CDC) operation to stop. The value can be either server time or commit time.

Server time example: --cdc-stop-position “server_time:2018-02-09T12:12:12”

Commit time example: --cdc-stop-position “commit_time: 2018-02-09T12:12:12 “

createReplicationTask_resourceIdentifier :: Lens' CreateReplicationTask (Maybe Text)

A friendly name for the resource identifier at the end of the EndpointArn response parameter that is returned in the created Endpoint object. The value for this parameter can have up to 31 characters. It can contain only ASCII letters, digits, and hyphen ('-'). Also, it can't end with a hyphen or contain two consecutive hyphens, and can only begin with a letter, such as Example-App-ARN1. For example, this value might result in the EndpointArn value arn:aws:dms:eu-west-1:012345678901:rep:Example-App-ARN1. If you don't specify a ResourceIdentifier value, DMS generates a default identifier value for the end of EndpointArn.

createReplicationTask_tags :: Lens' CreateReplicationTask (Maybe [Tag])

One or more tags to be assigned to the replication task.

createReplicationTask_cdcStartTime :: Lens' CreateReplicationTask (Maybe UTCTime)

Indicates the start time for a change data capture (CDC) operation. Use either CdcStartTime or CdcStartPosition to specify when you want a CDC operation to start. Specifying both values results in an error.

Timestamp Example: --cdc-start-time “2018-03-08T12:12:12”

createReplicationTask_replicationTaskIdentifier :: Lens' CreateReplicationTask Text

An identifier for the replication task.

Constraints:

Must contain 1-255 alphanumeric characters or hyphens.

First character must be a letter.

Cannot end with a hyphen or contain two consecutive hyphens.

createReplicationTask_sourceEndpointArn :: Lens' CreateReplicationTask Text

An Amazon Resource Name (ARN) that uniquely identifies the source endpoint.

createReplicationTask_targetEndpointArn :: Lens' CreateReplicationTask Text

An Amazon Resource Name (ARN) that uniquely identifies the target endpoint.

createReplicationTask_replicationInstanceArn :: Lens' CreateReplicationTask Text

The Amazon Resource Name (ARN) of a replication instance.

createReplicationTask_migrationType :: Lens' CreateReplicationTask MigrationTypeValue

The migration type. Valid values: full-load | cdc | full-load-and-cdc

createReplicationTask_tableMappings :: Lens' CreateReplicationTask Text

The table mappings for the task, in JSON format. For more information, see Using Table Mapping to Specify Task Settings in the Database Migration Service User Guide.

Destructuring the Response

data CreateReplicationTaskResponse

See: newCreateReplicationTaskResponse smart constructor.

Constructors

CreateReplicationTaskResponse'

Fields

replicationTask :: Maybe ReplicationTask

The replication task that was created.

httpStatus :: Int

The response's http status code.

newCreateReplicationTaskResponse

Arguments

:: Int
$sel:httpStatus:CreateReplicationTaskResponse'

-> CreateReplicationTaskResponse

Create a value of CreateReplicationTaskResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:replicationTask:CreateReplicationTaskResponse', createReplicationTaskResponse_replicationTask - The replication task that was created.

$sel:httpStatus:CreateReplicationTaskResponse', createReplicationTaskResponse_httpStatus - The response's http status code.

Response Lenses

createReplicationTaskResponse_replicationTask :: Lens' CreateReplicationTaskResponse (Maybe ReplicationTask)

The replication task that was created.

createReplicationTaskResponse_httpStatus :: Lens' CreateReplicationTaskResponse Int

The response's http status code.