Amazonka.DMS.DeleteEndpoint

Description

Deletes the specified endpoint.

All tasks associated with the endpoint must be deleted before you can delete the endpoint.

Creating a Request

data DeleteEndpoint

See: newDeleteEndpoint smart constructor.

Constructors

DeleteEndpoint'

Fields

endpointArn :: Text

The Amazon Resource Name (ARN) string that uniquely identifies the endpoint.

newDeleteEndpoint

Arguments

:: Text
$sel:endpointArn:DeleteEndpoint'

-> DeleteEndpoint

Create a value of DeleteEndpoint with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:endpointArn:DeleteEndpoint', deleteEndpoint_endpointArn - The Amazon Resource Name (ARN) string that uniquely identifies the endpoint.

Request Lenses

deleteEndpoint_endpointArn :: Lens' DeleteEndpoint Text

The Amazon Resource Name (ARN) string that uniquely identifies the endpoint.

Destructuring the Response

data DeleteEndpointResponse

See: newDeleteEndpointResponse smart constructor.

Constructors

DeleteEndpointResponse'

Fields

endpoint :: Maybe Endpoint

The endpoint that was deleted.

httpStatus :: Int

The response's http status code.

newDeleteEndpointResponse

Arguments

:: Int
$sel:httpStatus:DeleteEndpointResponse'

-> DeleteEndpointResponse

Create a value of DeleteEndpointResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:endpoint:DeleteEndpointResponse', deleteEndpointResponse_endpoint - The endpoint that was deleted.

$sel:httpStatus:DeleteEndpointResponse', deleteEndpointResponse_httpStatus - The response's http status code.

Response Lenses

deleteEndpointResponse_endpoint :: Lens' DeleteEndpointResponse (Maybe Endpoint)

The endpoint that was deleted.

deleteEndpointResponse_httpStatus :: Lens' DeleteEndpointResponse Int

The response's http status code.