Amazonka.DMS.DeleteReplicationTaskAssessmentRun

Description

Deletes the record of a single premigration assessment run.

This operation removes all metadata that DMS maintains about this assessment run. However, the operation leaves untouched all information about this assessment run that is stored in your Amazon S3 bucket.

Creating a Request

data DeleteReplicationTaskAssessmentRun

See: newDeleteReplicationTaskAssessmentRun smart constructor.

Constructors

DeleteReplicationTaskAssessmentRun'

Fields

replicationTaskAssessmentRunArn :: Text

Amazon Resource Name (ARN) of the premigration assessment run to be deleted.

newDeleteReplicationTaskAssessmentRun

Arguments

:: Text
$sel:replicationTaskAssessmentRunArn:DeleteReplicationTaskAssessmentRun'

-> DeleteReplicationTaskAssessmentRun

Create a value of DeleteReplicationTaskAssessmentRun with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:replicationTaskAssessmentRunArn:DeleteReplicationTaskAssessmentRun', deleteReplicationTaskAssessmentRun_replicationTaskAssessmentRunArn - Amazon Resource Name (ARN) of the premigration assessment run to be deleted.

Request Lenses

deleteReplicationTaskAssessmentRun_replicationTaskAssessmentRunArn :: Lens' DeleteReplicationTaskAssessmentRun Text

Amazon Resource Name (ARN) of the premigration assessment run to be deleted.

Destructuring the Response

data DeleteReplicationTaskAssessmentRunResponse

See: newDeleteReplicationTaskAssessmentRunResponse smart constructor.

Constructors

DeleteReplicationTaskAssessmentRunResponse'

Fields

replicationTaskAssessmentRun :: Maybe ReplicationTaskAssessmentRun

The ReplicationTaskAssessmentRun object for the deleted assessment run.

httpStatus :: Int

The response's http status code.

newDeleteReplicationTaskAssessmentRunResponse

Arguments

:: Int
$sel:httpStatus:DeleteReplicationTaskAssessmentRunResponse'

-> DeleteReplicationTaskAssessmentRunResponse

Create a value of DeleteReplicationTaskAssessmentRunResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:replicationTaskAssessmentRun:DeleteReplicationTaskAssessmentRunResponse', deleteReplicationTaskAssessmentRunResponse_replicationTaskAssessmentRun - The ReplicationTaskAssessmentRun object for the deleted assessment run.

$sel:httpStatus:DeleteReplicationTaskAssessmentRunResponse', deleteReplicationTaskAssessmentRunResponse_httpStatus - The response's http status code.

Response Lenses

deleteReplicationTaskAssessmentRunResponse_replicationTaskAssessmentRun :: Lens' DeleteReplicationTaskAssessmentRunResponse (Maybe ReplicationTaskAssessmentRun)

The ReplicationTaskAssessmentRun object for the deleted assessment run.

deleteReplicationTaskAssessmentRunResponse_httpStatus :: Lens' DeleteReplicationTaskAssessmentRunResponse Int

The response's http status code.