Amazonka.DMS.DescribeEventSubscriptions

Description

Lists all the event subscriptions for a customer account. The description of a subscription includes SubscriptionName, SNSTopicARN, CustomerID, SourceType, SourceID, CreationTime, and Status.

If you specify SubscriptionName, this action lists the description for that subscription.

This operation returns paginated results.

Creating a Request

data DescribeEventSubscriptions

See: newDescribeEventSubscriptions smart constructor.

Constructors

DescribeEventSubscriptions'

Fields

subscriptionName :: Maybe Text

The name of the DMS event subscription to be described.

filters :: Maybe [Filter]

Filters applied to event subscriptions.

marker :: Maybe Text

An optional pagination token provided by a previous request. If this parameter is specified, the response includes only records beyond the marker, up to the value specified by MaxRecords.

maxRecords :: Maybe Int

The maximum number of records to include in the response. If more records exist than the specified MaxRecords value, a pagination token called a marker is included in the response so that the remaining results can be retrieved.

Default: 100

Constraints: Minimum 20, maximum 100.

newDescribeEventSubscriptions :: DescribeEventSubscriptions

Create a value of DescribeEventSubscriptions with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:subscriptionName:DescribeEventSubscriptions', describeEventSubscriptions_subscriptionName - The name of the DMS event subscription to be described.

$sel:filters:DescribeEventSubscriptions', describeEventSubscriptions_filters - Filters applied to event subscriptions.

$sel:marker:DescribeEventSubscriptions', describeEventSubscriptions_marker - An optional pagination token provided by a previous request. If this parameter is specified, the response includes only records beyond the marker, up to the value specified by MaxRecords.

$sel:maxRecords:DescribeEventSubscriptions', describeEventSubscriptions_maxRecords - The maximum number of records to include in the response. If more records exist than the specified MaxRecords value, a pagination token called a marker is included in the response so that the remaining results can be retrieved.

Default: 100

Constraints: Minimum 20, maximum 100.

Request Lenses

describeEventSubscriptions_subscriptionName :: Lens' DescribeEventSubscriptions (Maybe Text)

The name of the DMS event subscription to be described.

describeEventSubscriptions_filters :: Lens' DescribeEventSubscriptions (Maybe [Filter])

Filters applied to event subscriptions.

describeEventSubscriptions_marker :: Lens' DescribeEventSubscriptions (Maybe Text)

An optional pagination token provided by a previous request. If this parameter is specified, the response includes only records beyond the marker, up to the value specified by MaxRecords.

describeEventSubscriptions_maxRecords :: Lens' DescribeEventSubscriptions (Maybe Int)

The maximum number of records to include in the response. If more records exist than the specified MaxRecords value, a pagination token called a marker is included in the response so that the remaining results can be retrieved.

Default: 100

Constraints: Minimum 20, maximum 100.

Destructuring the Response

data DescribeEventSubscriptionsResponse

See: newDescribeEventSubscriptionsResponse smart constructor.

Constructors

DescribeEventSubscriptionsResponse'

Fields

eventSubscriptionsList :: Maybe [EventSubscription]

A list of event subscriptions.

marker :: Maybe Text

An optional pagination token provided by a previous request. If this parameter is specified, the response includes only records beyond the marker, up to the value specified by MaxRecords.

httpStatus :: Int

The response's http status code.

newDescribeEventSubscriptionsResponse

Arguments

:: Int
$sel:httpStatus:DescribeEventSubscriptionsResponse'

-> DescribeEventSubscriptionsResponse

Create a value of DescribeEventSubscriptionsResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:eventSubscriptionsList:DescribeEventSubscriptionsResponse', describeEventSubscriptionsResponse_eventSubscriptionsList - A list of event subscriptions.

$sel:marker:DescribeEventSubscriptions', describeEventSubscriptionsResponse_marker - An optional pagination token provided by a previous request. If this parameter is specified, the response includes only records beyond the marker, up to the value specified by MaxRecords.

$sel:httpStatus:DescribeEventSubscriptionsResponse', describeEventSubscriptionsResponse_httpStatus - The response's http status code.

Response Lenses

describeEventSubscriptionsResponse_eventSubscriptionsList :: Lens' DescribeEventSubscriptionsResponse (Maybe [EventSubscription])

A list of event subscriptions.

describeEventSubscriptionsResponse_marker :: Lens' DescribeEventSubscriptionsResponse (Maybe Text)

An optional pagination token provided by a previous request. If this parameter is specified, the response includes only records beyond the marker, up to the value specified by MaxRecords.

describeEventSubscriptionsResponse_httpStatus :: Lens' DescribeEventSubscriptionsResponse Int

The response's http status code.