Amazonka.DMS.DescribeReplicationTaskIndividualAssessments

Description

Returns a paginated list of individual assessments based on filter settings.

These filter settings can specify a combination of premigration assessment runs, migration tasks, and assessment status values.

Creating a Request

data DescribeReplicationTaskIndividualAssessments

See: newDescribeReplicationTaskIndividualAssessments smart constructor.

Constructors

DescribeReplicationTaskIndividualAssessments'

Fields

filters :: Maybe [Filter]

Filters applied to the individual assessments described in the form of key-value pairs.

Valid filter names: replication-task-assessment-run-arn, replication-task-arn, status

marker :: Maybe Text

An optional pagination token provided by a previous request. If this parameter is specified, the response includes only records beyond the marker, up to the value specified by MaxRecords.

maxRecords :: Maybe Int

The maximum number of records to include in the response. If more records exist than the specified MaxRecords value, a pagination token called a marker is included in the response so that the remaining results can be retrieved.

newDescribeReplicationTaskIndividualAssessments :: DescribeReplicationTaskIndividualAssessments

Create a value of DescribeReplicationTaskIndividualAssessments with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:filters:DescribeReplicationTaskIndividualAssessments', describeReplicationTaskIndividualAssessments_filters - Filters applied to the individual assessments described in the form of key-value pairs.

Valid filter names: replication-task-assessment-run-arn, replication-task-arn, status

$sel:marker:DescribeReplicationTaskIndividualAssessments', describeReplicationTaskIndividualAssessments_marker - An optional pagination token provided by a previous request. If this parameter is specified, the response includes only records beyond the marker, up to the value specified by MaxRecords.

$sel:maxRecords:DescribeReplicationTaskIndividualAssessments', describeReplicationTaskIndividualAssessments_maxRecords - The maximum number of records to include in the response. If more records exist than the specified MaxRecords value, a pagination token called a marker is included in the response so that the remaining results can be retrieved.

Request Lenses

describeReplicationTaskIndividualAssessments_filters :: Lens' DescribeReplicationTaskIndividualAssessments (Maybe [Filter])

Filters applied to the individual assessments described in the form of key-value pairs.

Valid filter names: replication-task-assessment-run-arn, replication-task-arn, status

describeReplicationTaskIndividualAssessments_marker :: Lens' DescribeReplicationTaskIndividualAssessments (Maybe Text)

An optional pagination token provided by a previous request. If this parameter is specified, the response includes only records beyond the marker, up to the value specified by MaxRecords.

describeReplicationTaskIndividualAssessments_maxRecords :: Lens' DescribeReplicationTaskIndividualAssessments (Maybe Int)

The maximum number of records to include in the response. If more records exist than the specified MaxRecords value, a pagination token called a marker is included in the response so that the remaining results can be retrieved.

Destructuring the Response

data DescribeReplicationTaskIndividualAssessmentsResponse

See: newDescribeReplicationTaskIndividualAssessmentsResponse smart constructor.

Constructors

DescribeReplicationTaskIndividualAssessmentsResponse'

Fields

replicationTaskIndividualAssessments :: Maybe [ReplicationTaskIndividualAssessment]

One or more individual assessments as specified by Filters.

marker :: Maybe Text

A pagination token returned for you to pass to a subsequent request. If you pass this token as the Marker value in a subsequent request, the response includes only records beyond the marker, up to the value specified in the request by MaxRecords.

httpStatus :: Int

The response's http status code.

newDescribeReplicationTaskIndividualAssessmentsResponse

Arguments

:: Int
$sel:httpStatus:DescribeReplicationTaskIndividualAssessmentsResponse'

-> DescribeReplicationTaskIndividualAssessmentsResponse

Create a value of DescribeReplicationTaskIndividualAssessmentsResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:replicationTaskIndividualAssessments:DescribeReplicationTaskIndividualAssessmentsResponse', describeReplicationTaskIndividualAssessmentsResponse_replicationTaskIndividualAssessments - One or more individual assessments as specified by Filters.

$sel:marker:DescribeReplicationTaskIndividualAssessments', describeReplicationTaskIndividualAssessmentsResponse_marker - A pagination token returned for you to pass to a subsequent request. If you pass this token as the Marker value in a subsequent request, the response includes only records beyond the marker, up to the value specified in the request by MaxRecords.

$sel:httpStatus:DescribeReplicationTaskIndividualAssessmentsResponse', describeReplicationTaskIndividualAssessmentsResponse_httpStatus - The response's http status code.

Response Lenses

describeReplicationTaskIndividualAssessmentsResponse_replicationTaskIndividualAssessments :: Lens' DescribeReplicationTaskIndividualAssessmentsResponse (Maybe [ReplicationTaskIndividualAssessment])

One or more individual assessments as specified by Filters.

describeReplicationTaskIndividualAssessmentsResponse_marker :: Lens' DescribeReplicationTaskIndividualAssessmentsResponse (Maybe Text)

A pagination token returned for you to pass to a subsequent request. If you pass this token as the Marker value in a subsequent request, the response includes only records beyond the marker, up to the value specified in the request by MaxRecords.

describeReplicationTaskIndividualAssessmentsResponse_httpStatus :: Lens' DescribeReplicationTaskIndividualAssessmentsResponse Int

The response's http status code.