Amazonka.DMS.Types.Certificate

Description

Documentation

data Certificate

The SSL certificate that can be used to encrypt connections between the endpoints and the replication instance.

See: newCertificate smart constructor.

Constructors

Certificate'

Fields

certificateOwner :: Maybe Text

The owner of the certificate.

signingAlgorithm :: Maybe Text

The signing algorithm for the certificate.

validFromDate :: Maybe POSIX

The beginning date that the certificate is valid.

certificatePem :: Maybe Text

The contents of a .pem file, which contains an X.509 certificate.

certificateArn :: Maybe Text

The Amazon Resource Name (ARN) for the certificate.

certificateCreationDate :: Maybe POSIX

The date that the certificate was created.

certificateIdentifier :: Maybe Text

A customer-assigned name for the certificate. Identifiers must begin with a letter and must contain only ASCII letters, digits, and hyphens. They can't end with a hyphen or contain two consecutive hyphens.

certificateWallet :: Maybe Base64

The location of an imported Oracle Wallet certificate for use with SSL.

keyLength :: Maybe Int

The key length of the cryptographic algorithm being used.

validToDate :: Maybe POSIX

The final date that the certificate is valid.

newCertificate :: Certificate

Create a value of Certificate with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:certificateOwner:Certificate', certificate_certificateOwner - The owner of the certificate.

$sel:signingAlgorithm:Certificate', certificate_signingAlgorithm - The signing algorithm for the certificate.

$sel:validFromDate:Certificate', certificate_validFromDate - The beginning date that the certificate is valid.

$sel:certificatePem:Certificate', certificate_certificatePem - The contents of a .pem file, which contains an X.509 certificate.

$sel:certificateArn:Certificate', certificate_certificateArn - The Amazon Resource Name (ARN) for the certificate.

$sel:certificateCreationDate:Certificate', certificate_certificateCreationDate - The date that the certificate was created.

$sel:certificateIdentifier:Certificate', certificate_certificateIdentifier - A customer-assigned name for the certificate. Identifiers must begin with a letter and must contain only ASCII letters, digits, and hyphens. They can't end with a hyphen or contain two consecutive hyphens.

$sel:certificateWallet:Certificate', certificate_certificateWallet - The location of an imported Oracle Wallet certificate for use with SSL.-- -- Note: This Lens automatically encodes and decodes Base64 data. -- The underlying isomorphism will encode to Base64 representation during -- serialisation, and decode from Base64 representation during deserialisation. -- This Lens accepts and returns only raw unencoded data.

$sel:keyLength:Certificate', certificate_keyLength - The key length of the cryptographic algorithm being used.

$sel:validToDate:Certificate', certificate_validToDate - The final date that the certificate is valid.

certificate_certificateOwner :: Lens' Certificate (Maybe Text)

The owner of the certificate.

certificate_signingAlgorithm :: Lens' Certificate (Maybe Text)

The signing algorithm for the certificate.

certificate_validFromDate :: Lens' Certificate (Maybe UTCTime)

The beginning date that the certificate is valid.

certificate_certificatePem :: Lens' Certificate (Maybe Text)

The contents of a .pem file, which contains an X.509 certificate.

certificate_certificateArn :: Lens' Certificate (Maybe Text)

The Amazon Resource Name (ARN) for the certificate.

certificate_certificateCreationDate :: Lens' Certificate (Maybe UTCTime)

The date that the certificate was created.

certificate_certificateIdentifier :: Lens' Certificate (Maybe Text)

A customer-assigned name for the certificate. Identifiers must begin with a letter and must contain only ASCII letters, digits, and hyphens. They can't end with a hyphen or contain two consecutive hyphens.

certificate_certificateWallet :: Lens' Certificate (Maybe ByteString)

The location of an imported Oracle Wallet certificate for use with SSL.-- -- Note: This Lens automatically encodes and decodes Base64 data. -- The underlying isomorphism will encode to Base64 representation during -- serialisation, and decode from Base64 representation during deserialisation. -- This Lens accepts and returns only raw unencoded data.

certificate_keyLength :: Lens' Certificate (Maybe Int)

The key length of the cryptographic algorithm being used.

certificate_validToDate :: Lens' Certificate (Maybe UTCTime)

The final date that the certificate is valid.