Amazonka.DMS.Types.EndpointSetting

Description

Documentation

data EndpointSetting

Endpoint settings.

See: newEndpointSetting smart constructor.

Constructors

EndpointSetting'

Fields

sensitive :: Maybe Bool

A value that marks this endpoint setting as sensitive.

intValueMax :: Maybe Int

The maximum value of an endpoint setting that is of type int.

applicability :: Maybe Text

The relevance or validity of an endpoint setting for an engine name and its endpoint type.

name :: Maybe Text

The name that you want to give the endpoint settings.

intValueMin :: Maybe Int

The minimum value of an endpoint setting that is of type int.

units :: Maybe Text

The unit of measure for this endpoint setting.

defaultValue :: Maybe Text

The default value of the endpoint setting if no value is specified using CreateEndpoint or ModifyEndpoint.

type' :: Maybe EndpointSettingTypeValue

The type of endpoint. Valid values are source and target.

enumValues :: Maybe [Text]

Enumerated values to use for this endpoint.

newEndpointSetting :: EndpointSetting

Create a value of EndpointSetting with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:sensitive:EndpointSetting', endpointSetting_sensitive - A value that marks this endpoint setting as sensitive.

$sel:intValueMax:EndpointSetting', endpointSetting_intValueMax - The maximum value of an endpoint setting that is of type int.

$sel:applicability:EndpointSetting', endpointSetting_applicability - The relevance or validity of an endpoint setting for an engine name and its endpoint type.

$sel:name:EndpointSetting', endpointSetting_name - The name that you want to give the endpoint settings.

$sel:intValueMin:EndpointSetting', endpointSetting_intValueMin - The minimum value of an endpoint setting that is of type int.

$sel:units:EndpointSetting', endpointSetting_units - The unit of measure for this endpoint setting.

$sel:defaultValue:EndpointSetting', endpointSetting_defaultValue - The default value of the endpoint setting if no value is specified using CreateEndpoint or ModifyEndpoint.

$sel:type':EndpointSetting', endpointSetting_type - The type of endpoint. Valid values are source and target.

$sel:enumValues:EndpointSetting', endpointSetting_enumValues - Enumerated values to use for this endpoint.

endpointSetting_sensitive :: Lens' EndpointSetting (Maybe Bool)

A value that marks this endpoint setting as sensitive.

endpointSetting_intValueMax :: Lens' EndpointSetting (Maybe Int)

The maximum value of an endpoint setting that is of type int.

endpointSetting_applicability :: Lens' EndpointSetting (Maybe Text)

The relevance or validity of an endpoint setting for an engine name and its endpoint type.

endpointSetting_name :: Lens' EndpointSetting (Maybe Text)

The name that you want to give the endpoint settings.

endpointSetting_intValueMin :: Lens' EndpointSetting (Maybe Int)

The minimum value of an endpoint setting that is of type int.

endpointSetting_units :: Lens' EndpointSetting (Maybe Text)

The unit of measure for this endpoint setting.

endpointSetting_defaultValue :: Lens' EndpointSetting (Maybe Text)

The default value of the endpoint setting if no value is specified using CreateEndpoint or ModifyEndpoint.

endpointSetting_type :: Lens' EndpointSetting (Maybe EndpointSettingTypeValue)

The type of endpoint. Valid values are source and target.

endpointSetting_enumValues :: Lens' EndpointSetting (Maybe [Text])

Enumerated values to use for this endpoint.