Amazonka.DMS.Types.NeptuneSettings

Description

Documentation

data NeptuneSettings

Provides information that defines an Amazon Neptune endpoint.

See: newNeptuneSettings smart constructor.

Constructors

NeptuneSettings'

Fields

maxFileSize :: Maybe Int

The maximum size in kilobytes of migrated graph data stored in a .csv file before DMS bulk-loads the data to the Neptune target database. The default is 1,048,576 KB. If the bulk load is successful, DMS clears the bucket, ready to store the next batch of migrated graph data.

maxRetryCount :: Maybe Int

The number of times for DMS to retry a bulk load of migrated graph data to the Neptune target database before raising an error. The default is 5.

serviceAccessRoleArn :: Maybe Text

The Amazon Resource Name (ARN) of the service role that you created for the Neptune target endpoint. The role must allow the iam:PassRole action. For more information, see Creating an IAM Service Role for Accessing Amazon Neptune as a Target in the Database Migration Service User Guide.

iamAuthEnabled :: Maybe Bool

If you want Identity and Access Management (IAM) authorization enabled for this endpoint, set this parameter to true. Then attach the appropriate IAM policy document to your service role specified by ServiceAccessRoleArn. The default is false.

errorRetryDuration :: Maybe Int

The number of milliseconds for DMS to wait to retry a bulk-load of migrated graph data to the Neptune target database before raising an error. The default is 250.

s3BucketName :: Text

The name of the Amazon S3 bucket where DMS can temporarily store migrated graph data in .csv files before bulk-loading it to the Neptune target database. DMS maps the SQL source data to graph data before storing it in these .csv files.

s3BucketFolder :: Text

A folder path where you want DMS to store migrated graph data in the S3 bucket specified by S3BucketName

newNeptuneSettings

Arguments

:: Text
$sel:s3BucketName:NeptuneSettings'

-> Text
$sel:s3BucketFolder:NeptuneSettings'

-> NeptuneSettings

Create a value of NeptuneSettings with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:maxFileSize:NeptuneSettings', neptuneSettings_maxFileSize - The maximum size in kilobytes of migrated graph data stored in a .csv file before DMS bulk-loads the data to the Neptune target database. The default is 1,048,576 KB. If the bulk load is successful, DMS clears the bucket, ready to store the next batch of migrated graph data.

$sel:maxRetryCount:NeptuneSettings', neptuneSettings_maxRetryCount - The number of times for DMS to retry a bulk load of migrated graph data to the Neptune target database before raising an error. The default is 5.

$sel:serviceAccessRoleArn:NeptuneSettings', neptuneSettings_serviceAccessRoleArn - The Amazon Resource Name (ARN) of the service role that you created for the Neptune target endpoint. The role must allow the iam:PassRole action. For more information, see Creating an IAM Service Role for Accessing Amazon Neptune as a Target in the Database Migration Service User Guide.

$sel:iamAuthEnabled:NeptuneSettings', neptuneSettings_iamAuthEnabled - If you want Identity and Access Management (IAM) authorization enabled for this endpoint, set this parameter to true. Then attach the appropriate IAM policy document to your service role specified by ServiceAccessRoleArn. The default is false.

$sel:errorRetryDuration:NeptuneSettings', neptuneSettings_errorRetryDuration - The number of milliseconds for DMS to wait to retry a bulk-load of migrated graph data to the Neptune target database before raising an error. The default is 250.

$sel:s3BucketName:NeptuneSettings', neptuneSettings_s3BucketName - The name of the Amazon S3 bucket where DMS can temporarily store migrated graph data in .csv files before bulk-loading it to the Neptune target database. DMS maps the SQL source data to graph data before storing it in these .csv files.

$sel:s3BucketFolder:NeptuneSettings', neptuneSettings_s3BucketFolder - A folder path where you want DMS to store migrated graph data in the S3 bucket specified by S3BucketName

neptuneSettings_maxFileSize :: Lens' NeptuneSettings (Maybe Int)

The maximum size in kilobytes of migrated graph data stored in a .csv file before DMS bulk-loads the data to the Neptune target database. The default is 1,048,576 KB. If the bulk load is successful, DMS clears the bucket, ready to store the next batch of migrated graph data.

neptuneSettings_maxRetryCount :: Lens' NeptuneSettings (Maybe Int)

The number of times for DMS to retry a bulk load of migrated graph data to the Neptune target database before raising an error. The default is 5.

neptuneSettings_serviceAccessRoleArn :: Lens' NeptuneSettings (Maybe Text)

The Amazon Resource Name (ARN) of the service role that you created for the Neptune target endpoint. The role must allow the iam:PassRole action. For more information, see Creating an IAM Service Role for Accessing Amazon Neptune as a Target in the Database Migration Service User Guide.

neptuneSettings_iamAuthEnabled :: Lens' NeptuneSettings (Maybe Bool)

If you want Identity and Access Management (IAM) authorization enabled for this endpoint, set this parameter to true. Then attach the appropriate IAM policy document to your service role specified by ServiceAccessRoleArn. The default is false.

neptuneSettings_errorRetryDuration :: Lens' NeptuneSettings (Maybe Int)

The number of milliseconds for DMS to wait to retry a bulk-load of migrated graph data to the Neptune target database before raising an error. The default is 250.

neptuneSettings_s3BucketName :: Lens' NeptuneSettings Text

The name of the Amazon S3 bucket where DMS can temporarily store migrated graph data in .csv files before bulk-loading it to the Neptune target database. DMS maps the SQL source data to graph data before storing it in these .csv files.

neptuneSettings_s3BucketFolder :: Lens' NeptuneSettings Text

A folder path where you want DMS to store migrated graph data in the S3 bucket specified by S3BucketName