Amazonka.DMS.Types.RedisSettings

Description

Documentation

data RedisSettings

Provides information that defines a Redis target endpoint.

See: newRedisSettings smart constructor.

Constructors

RedisSettings'

Fields

sslSecurityProtocol :: Maybe SslSecurityProtocolValue

The connection to a Redis target endpoint using Transport Layer Security (TLS). Valid values include plaintext and ssl-encryption. The default is ssl-encryption. The ssl-encryption option makes an encrypted connection. Optionally, you can identify an Amazon Resource Name (ARN) for an SSL certificate authority (CA) using the SslCaCertificateArn setting. If an ARN isn't given for a CA, DMS uses the Amazon root CA.

The plaintext option doesn't provide Transport Layer Security (TLS) encryption for traffic between endpoint and database.

authUserName :: Maybe Text

The user name provided with the auth-role option of the AuthType setting for a Redis target endpoint.

sslCaCertificateArn :: Maybe Text

The Amazon Resource Name (ARN) for the certificate authority (CA) that DMS uses to connect to your Redis target endpoint.

authPassword :: Maybe (Sensitive Text)

The password provided with the auth-role and auth-token options of the AuthType setting for a Redis target endpoint.

authType :: Maybe RedisAuthTypeValue

The type of authentication to perform when connecting to a Redis target. Options include none, auth-token, and auth-role. The auth-token option requires an AuthPassword value to be provided. The auth-role option requires AuthUserName and AuthPassword values to be provided.

serverName :: Text

Fully qualified domain name of the endpoint.

port :: Int

Transmission Control Protocol (TCP) port for the endpoint.

newRedisSettings

Arguments

:: Text
$sel:serverName:RedisSettings'

-> Int
$sel:port:RedisSettings'

-> RedisSettings

Create a value of RedisSettings with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:sslSecurityProtocol:RedisSettings', redisSettings_sslSecurityProtocol - The connection to a Redis target endpoint using Transport Layer Security (TLS). Valid values include plaintext and ssl-encryption. The default is ssl-encryption. The ssl-encryption option makes an encrypted connection. Optionally, you can identify an Amazon Resource Name (ARN) for an SSL certificate authority (CA) using the SslCaCertificateArn setting. If an ARN isn't given for a CA, DMS uses the Amazon root CA.

The plaintext option doesn't provide Transport Layer Security (TLS) encryption for traffic between endpoint and database.

$sel:authUserName:RedisSettings', redisSettings_authUserName - The user name provided with the auth-role option of the AuthType setting for a Redis target endpoint.

$sel:sslCaCertificateArn:RedisSettings', redisSettings_sslCaCertificateArn - The Amazon Resource Name (ARN) for the certificate authority (CA) that DMS uses to connect to your Redis target endpoint.

$sel:authPassword:RedisSettings', redisSettings_authPassword - The password provided with the auth-role and auth-token options of the AuthType setting for a Redis target endpoint.

$sel:authType:RedisSettings', redisSettings_authType - The type of authentication to perform when connecting to a Redis target. Options include none, auth-token, and auth-role. The auth-token option requires an AuthPassword value to be provided. The auth-role option requires AuthUserName and AuthPassword values to be provided.

$sel:serverName:RedisSettings', redisSettings_serverName - Fully qualified domain name of the endpoint.

$sel:port:RedisSettings', redisSettings_port - Transmission Control Protocol (TCP) port for the endpoint.

redisSettings_sslSecurityProtocol :: Lens' RedisSettings (Maybe SslSecurityProtocolValue)

The connection to a Redis target endpoint using Transport Layer Security (TLS). Valid values include plaintext and ssl-encryption. The default is ssl-encryption. The ssl-encryption option makes an encrypted connection. Optionally, you can identify an Amazon Resource Name (ARN) for an SSL certificate authority (CA) using the SslCaCertificateArn setting. If an ARN isn't given for a CA, DMS uses the Amazon root CA.

The plaintext option doesn't provide Transport Layer Security (TLS) encryption for traffic between endpoint and database.

redisSettings_authUserName :: Lens' RedisSettings (Maybe Text)

The user name provided with the auth-role option of the AuthType setting for a Redis target endpoint.

redisSettings_sslCaCertificateArn :: Lens' RedisSettings (Maybe Text)

The Amazon Resource Name (ARN) for the certificate authority (CA) that DMS uses to connect to your Redis target endpoint.

redisSettings_authPassword :: Lens' RedisSettings (Maybe Text)

The password provided with the auth-role and auth-token options of the AuthType setting for a Redis target endpoint.

redisSettings_authType :: Lens' RedisSettings (Maybe RedisAuthTypeValue)

The type of authentication to perform when connecting to a Redis target. Options include none, auth-token, and auth-role. The auth-token option requires an AuthPassword value to be provided. The auth-role option requires AuthUserName and AuthPassword values to be provided.

redisSettings_serverName :: Lens' RedisSettings Text

Fully qualified domain name of the endpoint.

redisSettings_port :: Lens' RedisSettings Int

Transmission Control Protocol (TCP) port for the endpoint.