Amazonka.DMS.Types.ReplicationTaskAssessmentRun

Description

Documentation

data ReplicationTaskAssessmentRun

Provides information that describes a premigration assessment run that you have started using the StartReplicationTaskAssessmentRun operation.

Some of the information appears based on other operations that can return the ReplicationTaskAssessmentRun object.

See: newReplicationTaskAssessmentRun smart constructor.

Constructors

ReplicationTaskAssessmentRun'

Fields

status :: Maybe Text

Assessment run status.

This status can have one of the following values:

"cancelling" – The assessment run was canceled by the CancelReplicationTaskAssessmentRun operation.

"deleting" – The assessment run was deleted by the DeleteReplicationTaskAssessmentRun operation.

"failed" – At least one individual assessment completed with a failed status.

"error-provisioning" – An internal error occurred while resources were provisioned (during provisioning status).

"error-executing" – An internal error occurred while individual assessments ran (during running status).

"invalid state" – The assessment run is in an unknown state.

"passed" – All individual assessments have completed, and none has a failed status.

"provisioning" – Resources required to run individual assessments are being provisioned.

"running" – Individual assessments are being run.

"starting" – The assessment run is starting, but resources are not yet being provisioned for individual assessments.

serviceAccessRoleArn :: Maybe Text

ARN of the service role used to start the assessment run using the StartReplicationTaskAssessmentRun operation. The role must allow the iam:PassRole action.

replicationTaskAssessmentRunCreationDate :: Maybe POSIX

Date on which the assessment run was created using the StartReplicationTaskAssessmentRun operation.

assessmentProgress :: Maybe ReplicationTaskAssessmentRunProgress

Indication of the completion progress for the individual assessments specified to run.

resultKmsKeyArn :: Maybe Text

ARN of the KMS encryption key used to encrypt the assessment run results.

replicationTaskArn :: Maybe Text

ARN of the migration task associated with this premigration assessment run.

resultLocationBucket :: Maybe Text

Amazon S3 bucket where DMS stores the results of this assessment run.

resultLocationFolder :: Maybe Text

Folder in an Amazon S3 bucket where DMS stores the results of this assessment run.

resultEncryptionMode :: Maybe Text

Encryption mode used to encrypt the assessment run results.

assessmentRunName :: Maybe Text

Unique name of the assessment run.

replicationTaskAssessmentRunArn :: Maybe Text

Amazon Resource Name (ARN) of this assessment run.

lastFailureMessage :: Maybe Text

Last message generated by an individual assessment failure.

newReplicationTaskAssessmentRun :: ReplicationTaskAssessmentRun

Create a value of ReplicationTaskAssessmentRun with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:status:ReplicationTaskAssessmentRun', replicationTaskAssessmentRun_status - Assessment run status.

This status can have one of the following values:

"cancelling" – The assessment run was canceled by the CancelReplicationTaskAssessmentRun operation.

"deleting" – The assessment run was deleted by the DeleteReplicationTaskAssessmentRun operation.

"failed" – At least one individual assessment completed with a failed status.

"error-provisioning" – An internal error occurred while resources were provisioned (during provisioning status).

"error-executing" – An internal error occurred while individual assessments ran (during running status).

"invalid state" – The assessment run is in an unknown state.

"passed" – All individual assessments have completed, and none has a failed status.

"provisioning" – Resources required to run individual assessments are being provisioned.

"running" – Individual assessments are being run.

"starting" – The assessment run is starting, but resources are not yet being provisioned for individual assessments.

$sel:serviceAccessRoleArn:ReplicationTaskAssessmentRun', replicationTaskAssessmentRun_serviceAccessRoleArn - ARN of the service role used to start the assessment run using the StartReplicationTaskAssessmentRun operation. The role must allow the iam:PassRole action.

$sel:replicationTaskAssessmentRunCreationDate:ReplicationTaskAssessmentRun', replicationTaskAssessmentRun_replicationTaskAssessmentRunCreationDate - Date on which the assessment run was created using the StartReplicationTaskAssessmentRun operation.

$sel:assessmentProgress:ReplicationTaskAssessmentRun', replicationTaskAssessmentRun_assessmentProgress - Indication of the completion progress for the individual assessments specified to run.

$sel:resultKmsKeyArn:ReplicationTaskAssessmentRun', replicationTaskAssessmentRun_resultKmsKeyArn - ARN of the KMS encryption key used to encrypt the assessment run results.

$sel:replicationTaskArn:ReplicationTaskAssessmentRun', replicationTaskAssessmentRun_replicationTaskArn - ARN of the migration task associated with this premigration assessment run.

$sel:resultLocationBucket:ReplicationTaskAssessmentRun', replicationTaskAssessmentRun_resultLocationBucket - Amazon S3 bucket where DMS stores the results of this assessment run.

$sel:resultLocationFolder:ReplicationTaskAssessmentRun', replicationTaskAssessmentRun_resultLocationFolder - Folder in an Amazon S3 bucket where DMS stores the results of this assessment run.

$sel:resultEncryptionMode:ReplicationTaskAssessmentRun', replicationTaskAssessmentRun_resultEncryptionMode - Encryption mode used to encrypt the assessment run results.

$sel:assessmentRunName:ReplicationTaskAssessmentRun', replicationTaskAssessmentRun_assessmentRunName - Unique name of the assessment run.

$sel:replicationTaskAssessmentRunArn:ReplicationTaskAssessmentRun', replicationTaskAssessmentRun_replicationTaskAssessmentRunArn - Amazon Resource Name (ARN) of this assessment run.

$sel:lastFailureMessage:ReplicationTaskAssessmentRun', replicationTaskAssessmentRun_lastFailureMessage - Last message generated by an individual assessment failure.

replicationTaskAssessmentRun_status :: Lens' ReplicationTaskAssessmentRun (Maybe Text)

Assessment run status.

This status can have one of the following values:

"cancelling" – The assessment run was canceled by the CancelReplicationTaskAssessmentRun operation.

"deleting" – The assessment run was deleted by the DeleteReplicationTaskAssessmentRun operation.

"failed" – At least one individual assessment completed with a failed status.

"error-provisioning" – An internal error occurred while resources were provisioned (during provisioning status).

"error-executing" – An internal error occurred while individual assessments ran (during running status).

"invalid state" – The assessment run is in an unknown state.

"passed" – All individual assessments have completed, and none has a failed status.

"provisioning" – Resources required to run individual assessments are being provisioned.

"running" – Individual assessments are being run.

"starting" – The assessment run is starting, but resources are not yet being provisioned for individual assessments.

replicationTaskAssessmentRun_serviceAccessRoleArn :: Lens' ReplicationTaskAssessmentRun (Maybe Text)

ARN of the service role used to start the assessment run using the StartReplicationTaskAssessmentRun operation. The role must allow the iam:PassRole action.

replicationTaskAssessmentRun_replicationTaskAssessmentRunCreationDate :: Lens' ReplicationTaskAssessmentRun (Maybe UTCTime)

Date on which the assessment run was created using the StartReplicationTaskAssessmentRun operation.

replicationTaskAssessmentRun_assessmentProgress :: Lens' ReplicationTaskAssessmentRun (Maybe ReplicationTaskAssessmentRunProgress)

Indication of the completion progress for the individual assessments specified to run.

replicationTaskAssessmentRun_resultKmsKeyArn :: Lens' ReplicationTaskAssessmentRun (Maybe Text)

ARN of the KMS encryption key used to encrypt the assessment run results.

replicationTaskAssessmentRun_replicationTaskArn :: Lens' ReplicationTaskAssessmentRun (Maybe Text)

ARN of the migration task associated with this premigration assessment run.

replicationTaskAssessmentRun_resultLocationBucket :: Lens' ReplicationTaskAssessmentRun (Maybe Text)

Amazon S3 bucket where DMS stores the results of this assessment run.

replicationTaskAssessmentRun_resultLocationFolder :: Lens' ReplicationTaskAssessmentRun (Maybe Text)

Folder in an Amazon S3 bucket where DMS stores the results of this assessment run.

replicationTaskAssessmentRun_resultEncryptionMode :: Lens' ReplicationTaskAssessmentRun (Maybe Text)

Encryption mode used to encrypt the assessment run results.

replicationTaskAssessmentRun_assessmentRunName :: Lens' ReplicationTaskAssessmentRun (Maybe Text)

Unique name of the assessment run.

replicationTaskAssessmentRun_replicationTaskAssessmentRunArn :: Lens' ReplicationTaskAssessmentRun (Maybe Text)

Amazon Resource Name (ARN) of this assessment run.

replicationTaskAssessmentRun_lastFailureMessage :: Lens' ReplicationTaskAssessmentRun (Maybe Text)

Last message generated by an individual assessment failure.