Amazonka.DMS.Types.TableStatistics

Description

Documentation

data TableStatistics

Provides a collection of table statistics in response to a request by the DescribeTableStatistics operation.

See: newTableStatistics smart constructor.

Constructors

TableStatistics'

Fields

validationState :: Maybe Text

The validation state of the table.

This parameter can have the following values:

Not enabled – Validation isn't enabled for the table in the migration task.

Pending records – Some records in the table are waiting for validation.

Mismatched records – Some records in the table don't match between the source and target.

Suspended records – Some records in the table couldn't be validated.

No primary key –The table couldn't be validated because it has no primary key.

Table error – The table wasn't validated because it's in an error state and some data wasn't migrated.

Validated – All rows in the table are validated. If the table is updated, the status can change from Validated.

Error – The table couldn't be validated because of an unexpected error.

Pending validation – The table is waiting validation.

Preparing table – Preparing the table enabled in the migration task for validation.

Pending revalidation – All rows in the table are pending validation after the table was updated.

fullLoadRows :: Maybe Integer

The number of rows added during the full load operation.

inserts :: Maybe Integer

The number of insert actions performed on a table.

fullLoadEndTime :: Maybe POSIX

The time when the full load operation completed.

fullLoadCondtnlChkFailedRows :: Maybe Integer

The number of rows that failed conditional checks during the full load operation (valid only for migrations where DynamoDB is the target).

fullLoadReloaded :: Maybe Bool

A value that indicates if the table was reloaded (true) or loaded as part of a new full load operation (false).

validationFailedRecords :: Maybe Integer

The number of records that failed validation.

validationSuspendedRecords :: Maybe Integer

The number of records that couldn't be validated.

schemaName :: Maybe Text

The schema name.

validationStateDetails :: Maybe Text

Additional details about the state of validation.

tableState :: Maybe Text

The state of the tables described.

Valid states: Table does not exist | Before load | Full load | Table completed | Table cancelled | Table error | Table all | Table updates | Table is being reloaded

fullLoadErrorRows :: Maybe Integer

The number of rows that failed to load during the full load operation (valid only for migrations where DynamoDB is the target).

ddls :: Maybe Integer

The data definition language (DDL) used to build and modify the structure of your tables.

deletes :: Maybe Integer

The number of delete actions performed on a table.

updates :: Maybe Integer

The number of update actions performed on a table.

validationPendingRecords :: Maybe Integer

The number of records that have yet to be validated.

fullLoadStartTime :: Maybe POSIX

The time when the full load operation started.

lastUpdateTime :: Maybe POSIX

The last time a table was updated.

tableName :: Maybe Text

The name of the table.

newTableStatistics :: TableStatistics

Create a value of TableStatistics with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:validationState:TableStatistics', tableStatistics_validationState - The validation state of the table.

This parameter can have the following values:

Not enabled – Validation isn't enabled for the table in the migration task.

Pending records – Some records in the table are waiting for validation.

Mismatched records – Some records in the table don't match between the source and target.

Suspended records – Some records in the table couldn't be validated.

No primary key –The table couldn't be validated because it has no primary key.

Table error – The table wasn't validated because it's in an error state and some data wasn't migrated.

Validated – All rows in the table are validated. If the table is updated, the status can change from Validated.

Error – The table couldn't be validated because of an unexpected error.

Pending validation – The table is waiting validation.

Preparing table – Preparing the table enabled in the migration task for validation.

Pending revalidation – All rows in the table are pending validation after the table was updated.

$sel:fullLoadRows:TableStatistics', tableStatistics_fullLoadRows - The number of rows added during the full load operation.

$sel:inserts:TableStatistics', tableStatistics_inserts - The number of insert actions performed on a table.

$sel:fullLoadEndTime:TableStatistics', tableStatistics_fullLoadEndTime - The time when the full load operation completed.

$sel:fullLoadCondtnlChkFailedRows:TableStatistics', tableStatistics_fullLoadCondtnlChkFailedRows - The number of rows that failed conditional checks during the full load operation (valid only for migrations where DynamoDB is the target).

$sel:fullLoadReloaded:TableStatistics', tableStatistics_fullLoadReloaded - A value that indicates if the table was reloaded (true) or loaded as part of a new full load operation (false).

$sel:validationFailedRecords:TableStatistics', tableStatistics_validationFailedRecords - The number of records that failed validation.

$sel:validationSuspendedRecords:TableStatistics', tableStatistics_validationSuspendedRecords - The number of records that couldn't be validated.

$sel:schemaName:TableStatistics', tableStatistics_schemaName - The schema name.

$sel:validationStateDetails:TableStatistics', tableStatistics_validationStateDetails - Additional details about the state of validation.

$sel:tableState:TableStatistics', tableStatistics_tableState - The state of the tables described.

Valid states: Table does not exist | Before load | Full load | Table completed | Table cancelled | Table error | Table all | Table updates | Table is being reloaded

$sel:fullLoadErrorRows:TableStatistics', tableStatistics_fullLoadErrorRows - The number of rows that failed to load during the full load operation (valid only for migrations where DynamoDB is the target).

$sel:ddls:TableStatistics', tableStatistics_ddls - The data definition language (DDL) used to build and modify the structure of your tables.

$sel:deletes:TableStatistics', tableStatistics_deletes - The number of delete actions performed on a table.

$sel:updates:TableStatistics', tableStatistics_updates - The number of update actions performed on a table.

$sel:validationPendingRecords:TableStatistics', tableStatistics_validationPendingRecords - The number of records that have yet to be validated.

$sel:fullLoadStartTime:TableStatistics', tableStatistics_fullLoadStartTime - The time when the full load operation started.

$sel:lastUpdateTime:TableStatistics', tableStatistics_lastUpdateTime - The last time a table was updated.

$sel:tableName:TableStatistics', tableStatistics_tableName - The name of the table.

tableStatistics_validationState :: Lens' TableStatistics (Maybe Text)

The validation state of the table.

This parameter can have the following values:

Not enabled – Validation isn't enabled for the table in the migration task.

Pending records – Some records in the table are waiting for validation.

Mismatched records – Some records in the table don't match between the source and target.

Suspended records – Some records in the table couldn't be validated.

No primary key –The table couldn't be validated because it has no primary key.

Table error – The table wasn't validated because it's in an error state and some data wasn't migrated.

Validated – All rows in the table are validated. If the table is updated, the status can change from Validated.

Error – The table couldn't be validated because of an unexpected error.

Pending validation – The table is waiting validation.

Preparing table – Preparing the table enabled in the migration task for validation.

Pending revalidation – All rows in the table are pending validation after the table was updated.

tableStatistics_fullLoadRows :: Lens' TableStatistics (Maybe Integer)

The number of rows added during the full load operation.

tableStatistics_inserts :: Lens' TableStatistics (Maybe Integer)

The number of insert actions performed on a table.

tableStatistics_fullLoadEndTime :: Lens' TableStatistics (Maybe UTCTime)

The time when the full load operation completed.

tableStatistics_fullLoadCondtnlChkFailedRows :: Lens' TableStatistics (Maybe Integer)

The number of rows that failed conditional checks during the full load operation (valid only for migrations where DynamoDB is the target).

tableStatistics_fullLoadReloaded :: Lens' TableStatistics (Maybe Bool)

A value that indicates if the table was reloaded (true) or loaded as part of a new full load operation (false).

tableStatistics_validationFailedRecords :: Lens' TableStatistics (Maybe Integer)

The number of records that failed validation.

tableStatistics_validationSuspendedRecords :: Lens' TableStatistics (Maybe Integer)

The number of records that couldn't be validated.

tableStatistics_schemaName :: Lens' TableStatistics (Maybe Text)

The schema name.

tableStatistics_validationStateDetails :: Lens' TableStatistics (Maybe Text)

Additional details about the state of validation.

tableStatistics_tableState :: Lens' TableStatistics (Maybe Text)

The state of the tables described.

Valid states: Table does not exist | Before load | Full load | Table completed | Table cancelled | Table error | Table all | Table updates | Table is being reloaded

tableStatistics_fullLoadErrorRows :: Lens' TableStatistics (Maybe Integer)

The number of rows that failed to load during the full load operation (valid only for migrations where DynamoDB is the target).

tableStatistics_ddls :: Lens' TableStatistics (Maybe Integer)

The data definition language (DDL) used to build and modify the structure of your tables.

tableStatistics_deletes :: Lens' TableStatistics (Maybe Integer)

The number of delete actions performed on a table.

tableStatistics_updates :: Lens' TableStatistics (Maybe Integer)

The number of update actions performed on a table.

tableStatistics_validationPendingRecords :: Lens' TableStatistics (Maybe Integer)

The number of records that have yet to be validated.

tableStatistics_fullLoadStartTime :: Lens' TableStatistics (Maybe UTCTime)

The time when the full load operation started.

tableStatistics_lastUpdateTime :: Lens' TableStatistics (Maybe UTCTime)

The last time a table was updated.

tableStatistics_tableName :: Lens' TableStatistics (Maybe Text)

The name of the table.