Amazonka.Data.Time

Description

Time

data Format

Constructors

RFC822Format

ISO8601Format

BasicFormat

AWSFormat

POSIXFormat

newtype Time (a :: Format)

Constructors

Time

Fields

fromTime :: UTCTime

_Time :: Iso' (Time a) UTCTime

Formats

data UTCTime

This is the simplest representation of UTC. It consists of the day number, and a time offset from midnight. Note that if a day has a leap second added to it, it will have 86401 seconds.

type RFC822 = Time 'RFC822Format

type ISO8601 = Time 'ISO8601Format

type BasicTime = Time 'BasicFormat

type AWSTime = Time 'AWSFormat

type POSIX = Time 'POSIXFormat