Amazonka.Data.XML

Description

Documentation

(.@) :: FromXML a => [Node] -> Text -> Either String a infixl 7

(.@?) :: FromXML a => [Node] -> Text -> Either String (Maybe a) infixl 7

(@=) :: ToXML a => Name -> a -> XML infixr 7

(@@=) :: ToText a => Name -> a -> XML infixr 7

decodeXML :: FromXML a => ByteStringLazy -> Either String a

encodeXML :: ToElement a => a -> ByteStringLazy

class FromXML a where

Methods

parseXML :: [Node] -> Either String a

class ToElement a where

Methods

toElement :: a -> Element

maybeElement :: ToElement a => a -> Maybe Element

Convert to an Element, only if the resulting element contains > 0 nodes.

data XML

Provides a way to make the operators for ToXML instance declaration be consistent WRT to single nodes or lists of nodes.

Constructors

XNull

XAttr Name Text

XOne Node

XMany [(Name, Text)] [Node]

listXMLNodes :: XML -> [Node]

listXMLAttributes :: XML -> [(Name, Text)]

class ToXML a where

Methods

toXML :: a -> XML

parseXMLMap :: (Eq k, Hashable k, FromText k, FromXML v) => Text -> Text -> Text -> [Node] -> Either String (HashMap k v)

parseXMLList1 :: FromXML a => Text -> [Node] -> Either String (NonEmpty a)

parseXMLList :: FromXML a => Text -> [Node] -> Either String [a]

parseXMLText :: FromText a => String -> [Node] -> Either String a

toXMLList :: (IsList a, ToXML (Item a)) => Name -> a -> XML

toXMLText :: ToText a => a -> XML

mkElement :: ToXML a => Name -> a -> Element

withContent :: String -> [Node] -> Either String (Maybe Text)

findElement :: Text -> [Node] -> Either String [Node]

Find a specific named NodeElement, at the current depth in the node tree.

Fails if absent.

firstElement :: Text -> [Node] -> Either String [Node]

Find the first specific named NodeElement, at any depth in the node tree.

Fails if absent.

childNodesOf :: Text -> Node -> Maybe [Node]

childrenOf :: Element -> [Node]

localName :: Node -> Maybe Text

rootElementName :: ByteStringLazy -> Maybe Text

An inefficient mechanism for retreiving the root element name of an XML document.

missingElement :: Text -> [Node] -> Maybe a -> Either String a