Amazonka.DataBrew.CreateProfileJob

Description

Creates a new job to analyze a dataset and create its data profile.

Creating a Request

data CreateProfileJob

See: newCreateProfileJob smart constructor.

Constructors

CreateProfileJob'

Fields

encryptionMode :: Maybe EncryptionMode

The encryption mode for the job, which can be one of the following:

SSE-KMS - SSE-KMS - Server-side encryption with KMS-managed keys.

SSE-S3 - Server-side encryption with keys managed by Amazon S3.

logSubscription :: Maybe LogSubscription

Enables or disables Amazon CloudWatch logging for the job. If logging is enabled, CloudWatch writes one log stream for each job run.

maxRetries :: Maybe Natural

The maximum number of times to retry the job after a job run fails.

encryptionKeyArn :: Maybe Text

The Amazon Resource Name (ARN) of an encryption key that is used to protect the job.

maxCapacity :: Maybe Int

The maximum number of nodes that DataBrew can use when the job processes data.

configuration :: Maybe ProfileConfiguration

Configuration for profile jobs. Used to select columns, do evaluations, and override default parameters of evaluations. When configuration is null, the profile job will run with default settings.

timeout :: Maybe Natural

The job's timeout in minutes. A job that attempts to run longer than this timeout period ends with a status of TIMEOUT.

tags :: Maybe (HashMap Text Text)

Metadata tags to apply to this job.

jobSample :: Maybe JobSample

Sample configuration for profile jobs only. Determines the number of rows on which the profile job will be executed. If a JobSample value is not provided, the default value will be used. The default value is CUSTOM_ROWS for the mode parameter and 20000 for the size parameter.

datasetName :: Text

The name of the dataset that this job is to act upon.

name :: Text

The name of the job to be created. Valid characters are alphanumeric (A-Z, a-z, 0-9), hyphen (-), period (.), and space.

outputLocation :: S3Location

roleArn :: Text

The Amazon Resource Name (ARN) of the Identity and Access Management (IAM) role to be assumed when DataBrew runs the job.

newCreateProfileJob

Arguments

:: Text
$sel:datasetName:CreateProfileJob'

-> Text
$sel:name:CreateProfileJob'

-> S3Location
$sel:outputLocation:CreateProfileJob'

-> Text
$sel:roleArn:CreateProfileJob'

-> CreateProfileJob

Create a value of CreateProfileJob with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:encryptionMode:CreateProfileJob', createProfileJob_encryptionMode - The encryption mode for the job, which can be one of the following:

SSE-KMS - SSE-KMS - Server-side encryption with KMS-managed keys.

SSE-S3 - Server-side encryption with keys managed by Amazon S3.

$sel:logSubscription:CreateProfileJob', createProfileJob_logSubscription - Enables or disables Amazon CloudWatch logging for the job. If logging is enabled, CloudWatch writes one log stream for each job run.

$sel:maxRetries:CreateProfileJob', createProfileJob_maxRetries - The maximum number of times to retry the job after a job run fails.

$sel:encryptionKeyArn:CreateProfileJob', createProfileJob_encryptionKeyArn - The Amazon Resource Name (ARN) of an encryption key that is used to protect the job.

$sel:maxCapacity:CreateProfileJob', createProfileJob_maxCapacity - The maximum number of nodes that DataBrew can use when the job processes data.

$sel:configuration:CreateProfileJob', createProfileJob_configuration - Configuration for profile jobs. Used to select columns, do evaluations, and override default parameters of evaluations. When configuration is null, the profile job will run with default settings.

$sel:timeout:CreateProfileJob', createProfileJob_timeout - The job's timeout in minutes. A job that attempts to run longer than this timeout period ends with a status of TIMEOUT.

$sel:tags:CreateProfileJob', createProfileJob_tags - Metadata tags to apply to this job.

$sel:jobSample:CreateProfileJob', createProfileJob_jobSample - Sample configuration for profile jobs only. Determines the number of rows on which the profile job will be executed. If a JobSample value is not provided, the default value will be used. The default value is CUSTOM_ROWS for the mode parameter and 20000 for the size parameter.

$sel:datasetName:CreateProfileJob', createProfileJob_datasetName - The name of the dataset that this job is to act upon.

$sel:name:CreateProfileJob', createProfileJob_name - The name of the job to be created. Valid characters are alphanumeric (A-Z, a-z, 0-9), hyphen (-), period (.), and space.

$sel:outputLocation:CreateProfileJob', createProfileJob_outputLocation - Undocumented member.

$sel:roleArn:CreateProfileJob', createProfileJob_roleArn - The Amazon Resource Name (ARN) of the Identity and Access Management (IAM) role to be assumed when DataBrew runs the job.

Request Lenses

createProfileJob_encryptionMode :: Lens' CreateProfileJob (Maybe EncryptionMode)

The encryption mode for the job, which can be one of the following:

SSE-KMS - SSE-KMS - Server-side encryption with KMS-managed keys.

SSE-S3 - Server-side encryption with keys managed by Amazon S3.

createProfileJob_logSubscription :: Lens' CreateProfileJob (Maybe LogSubscription)

Enables or disables Amazon CloudWatch logging for the job. If logging is enabled, CloudWatch writes one log stream for each job run.

createProfileJob_maxRetries :: Lens' CreateProfileJob (Maybe Natural)

The maximum number of times to retry the job after a job run fails.

createProfileJob_encryptionKeyArn :: Lens' CreateProfileJob (Maybe Text)

The Amazon Resource Name (ARN) of an encryption key that is used to protect the job.

createProfileJob_maxCapacity :: Lens' CreateProfileJob (Maybe Int)

The maximum number of nodes that DataBrew can use when the job processes data.

createProfileJob_configuration :: Lens' CreateProfileJob (Maybe ProfileConfiguration)

Configuration for profile jobs. Used to select columns, do evaluations, and override default parameters of evaluations. When configuration is null, the profile job will run with default settings.

createProfileJob_timeout :: Lens' CreateProfileJob (Maybe Natural)

The job's timeout in minutes. A job that attempts to run longer than this timeout period ends with a status of TIMEOUT.

createProfileJob_tags :: Lens' CreateProfileJob (Maybe (HashMap Text Text))

Metadata tags to apply to this job.

createProfileJob_jobSample :: Lens' CreateProfileJob (Maybe JobSample)

Sample configuration for profile jobs only. Determines the number of rows on which the profile job will be executed. If a JobSample value is not provided, the default value will be used. The default value is CUSTOM_ROWS for the mode parameter and 20000 for the size parameter.

createProfileJob_datasetName :: Lens' CreateProfileJob Text

The name of the dataset that this job is to act upon.

createProfileJob_name :: Lens' CreateProfileJob Text

The name of the job to be created. Valid characters are alphanumeric (A-Z, a-z, 0-9), hyphen (-), period (.), and space.

createProfileJob_outputLocation :: Lens' CreateProfileJob S3Location

Undocumented member.

createProfileJob_roleArn :: Lens' CreateProfileJob Text

The Amazon Resource Name (ARN) of the Identity and Access Management (IAM) role to be assumed when DataBrew runs the job.

Destructuring the Response

data CreateProfileJobResponse

See: newCreateProfileJobResponse smart constructor.

Constructors

CreateProfileJobResponse'

Fields

httpStatus :: Int

The response's http status code.

name :: Text

The name of the job that was created.

newCreateProfileJobResponse

Arguments

:: Int
$sel:httpStatus:CreateProfileJobResponse'

-> Text
$sel:name:CreateProfileJob'

-> CreateProfileJobResponse

Create a value of CreateProfileJobResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:httpStatus:CreateProfileJobResponse', createProfileJobResponse_httpStatus - The response's http status code.

$sel:name:CreateProfileJob', createProfileJobResponse_name - The name of the job that was created.

Response Lenses

createProfileJobResponse_httpStatus :: Lens' CreateProfileJobResponse Int

The response's http status code.

createProfileJobResponse_name :: Lens' CreateProfileJobResponse Text

The name of the job that was created.