Amazonka.DataBrew.DescribeDataset

Description

Returns the definition of a specific DataBrew dataset.

Creating a Request

data DescribeDataset

See: newDescribeDataset smart constructor.

Constructors

DescribeDataset'

Fields

name :: Text

The name of the dataset to be described.

newDescribeDataset

Arguments

:: Text
$sel:name:DescribeDataset'

-> DescribeDataset

Create a value of DescribeDataset with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:name:DescribeDataset', describeDataset_name - The name of the dataset to be described.

Request Lenses

describeDataset_name :: Lens' DescribeDataset Text

The name of the dataset to be described.

Destructuring the Response

data DescribeDatasetResponse

See: newDescribeDatasetResponse smart constructor.

Constructors

DescribeDatasetResponse'

Fields

lastModifiedDate :: Maybe POSIX

The date and time that the dataset was last modified.

pathOptions :: Maybe PathOptions

A set of options that defines how DataBrew interprets an Amazon S3 path of the dataset.

createDate :: Maybe POSIX

The date and time that the dataset was created.

formatOptions :: Maybe FormatOptions

format :: Maybe InputFormat

The file format of a dataset that is created from an Amazon S3 file or folder.

createdBy :: Maybe Text

The identifier (user name) of the user who created the dataset.

resourceArn :: Maybe Text

The Amazon Resource Name (ARN) of the dataset.

source :: Maybe Source

The location of the data for this dataset, Amazon S3 or the Glue Data Catalog.

lastModifiedBy :: Maybe Text

The identifier (user name) of the user who last modified the dataset.

tags :: Maybe (HashMap Text Text)

Metadata tags associated with this dataset.

httpStatus :: Int

The response's http status code.

name :: Text

The name of the dataset.

input :: Input

newDescribeDatasetResponse

Arguments

:: Int
$sel:httpStatus:DescribeDatasetResponse'

-> Text
$sel:name:DescribeDataset'

-> Input
$sel:input:DescribeDatasetResponse'

-> DescribeDatasetResponse

Create a value of DescribeDatasetResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:lastModifiedDate:DescribeDatasetResponse', describeDatasetResponse_lastModifiedDate - The date and time that the dataset was last modified.

$sel:pathOptions:DescribeDatasetResponse', describeDatasetResponse_pathOptions - A set of options that defines how DataBrew interprets an Amazon S3 path of the dataset.

$sel:createDate:DescribeDatasetResponse', describeDatasetResponse_createDate - The date and time that the dataset was created.

$sel:formatOptions:DescribeDatasetResponse', describeDatasetResponse_formatOptions - Undocumented member.

$sel:format:DescribeDatasetResponse', describeDatasetResponse_format - The file format of a dataset that is created from an Amazon S3 file or folder.

$sel:createdBy:DescribeDatasetResponse', describeDatasetResponse_createdBy - The identifier (user name) of the user who created the dataset.

$sel:resourceArn:DescribeDatasetResponse', describeDatasetResponse_resourceArn - The Amazon Resource Name (ARN) of the dataset.

$sel:source:DescribeDatasetResponse', describeDatasetResponse_source - The location of the data for this dataset, Amazon S3 or the Glue Data Catalog.

$sel:lastModifiedBy:DescribeDatasetResponse', describeDatasetResponse_lastModifiedBy - The identifier (user name) of the user who last modified the dataset.

$sel:tags:DescribeDatasetResponse', describeDatasetResponse_tags - Metadata tags associated with this dataset.

$sel:httpStatus:DescribeDatasetResponse', describeDatasetResponse_httpStatus - The response's http status code.

$sel:name:DescribeDataset', describeDatasetResponse_name - The name of the dataset.

$sel:input:DescribeDatasetResponse', describeDatasetResponse_input - Undocumented member.

Response Lenses

describeDatasetResponse_lastModifiedDate :: Lens' DescribeDatasetResponse (Maybe UTCTime)

The date and time that the dataset was last modified.

describeDatasetResponse_pathOptions :: Lens' DescribeDatasetResponse (Maybe PathOptions)

A set of options that defines how DataBrew interprets an Amazon S3 path of the dataset.

describeDatasetResponse_createDate :: Lens' DescribeDatasetResponse (Maybe UTCTime)

The date and time that the dataset was created.

describeDatasetResponse_formatOptions :: Lens' DescribeDatasetResponse (Maybe FormatOptions)

Undocumented member.

describeDatasetResponse_format :: Lens' DescribeDatasetResponse (Maybe InputFormat)

The file format of a dataset that is created from an Amazon S3 file or folder.

describeDatasetResponse_createdBy :: Lens' DescribeDatasetResponse (Maybe Text)

The identifier (user name) of the user who created the dataset.

describeDatasetResponse_resourceArn :: Lens' DescribeDatasetResponse (Maybe Text)

The Amazon Resource Name (ARN) of the dataset.

describeDatasetResponse_source :: Lens' DescribeDatasetResponse (Maybe Source)

The location of the data for this dataset, Amazon S3 or the Glue Data Catalog.

describeDatasetResponse_lastModifiedBy :: Lens' DescribeDatasetResponse (Maybe Text)

The identifier (user name) of the user who last modified the dataset.

describeDatasetResponse_tags :: Lens' DescribeDatasetResponse (Maybe (HashMap Text Text))

Metadata tags associated with this dataset.

describeDatasetResponse_httpStatus :: Lens' DescribeDatasetResponse Int

The response's http status code.

describeDatasetResponse_name :: Lens' DescribeDatasetResponse Text

The name of the dataset.

describeDatasetResponse_input :: Lens' DescribeDatasetResponse Input

Undocumented member.