Amazonka.DataBrew.DescribeJobRun

Description

Represents one run of a DataBrew job.

Creating a Request

data DescribeJobRun

See: newDescribeJobRun smart constructor.

Constructors

DescribeJobRun'

Fields

name :: Text

The name of the job being processed during this run.

runId :: Text

The unique identifier of the job run.

newDescribeJobRun

Arguments

:: Text
$sel:name:DescribeJobRun'

-> Text
$sel:runId:DescribeJobRun'

-> DescribeJobRun

Create a value of DescribeJobRun with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:name:DescribeJobRun', describeJobRun_name - The name of the job being processed during this run.

$sel:runId:DescribeJobRun', describeJobRun_runId - The unique identifier of the job run.

Request Lenses

describeJobRun_name :: Lens' DescribeJobRun Text

The name of the job being processed during this run.

describeJobRun_runId :: Lens' DescribeJobRun Text

The unique identifier of the job run.

Destructuring the Response

data DescribeJobRunResponse

See: newDescribeJobRunResponse smart constructor.

Constructors

DescribeJobRunResponse'

Fields

completedOn :: Maybe POSIX

The date and time when the job completed processing.

state :: Maybe JobRunState

The current state of the job run entity itself.

dataCatalogOutputs :: Maybe (NonEmpty DataCatalogOutput)

One or more artifacts that represent the Glue Data Catalog output from running the job.

startedOn :: Maybe POSIX

The date and time when the job run began.

recipeReference :: Maybe RecipeReference

profileConfiguration :: Maybe ProfileConfiguration

Configuration for profile jobs. Used to select columns, do evaluations, and override default parameters of evaluations. When configuration is null, the profile job will run with default settings.

databaseOutputs :: Maybe (NonEmpty DatabaseOutput)

Represents a list of JDBC database output objects which defines the output destination for a DataBrew recipe job to write into.

logGroupName :: Maybe Text

The name of an Amazon CloudWatch log group, where the job writes diagnostic messages when it runs.

outputs :: Maybe (NonEmpty Output)

One or more output artifacts from a job run.

runId :: Maybe Text

The unique identifier of the job run.

executionTime :: Maybe Int

The amount of time, in seconds, during which the job run consumed resources.

logSubscription :: Maybe LogSubscription

The current status of Amazon CloudWatch logging for the job run.

startedBy :: Maybe Text

The Amazon Resource Name (ARN) of the user who started the job run.

datasetName :: Maybe Text

The name of the dataset for the job to process.

attempt :: Maybe Int

The number of times that DataBrew has attempted to run the job.

errorMessage :: Maybe Text

A message indicating an error (if any) that was encountered when the job ran.

jobSample :: Maybe JobSample

Sample configuration for profile jobs only. Determines the number of rows on which the profile job will be executed. If a JobSample value is not provided, the default value will be used. The default value is CUSTOM_ROWS for the mode parameter and 20000 for the size parameter.

httpStatus :: Int

The response's http status code.

jobName :: Text

The name of the job being processed during this run.

newDescribeJobRunResponse

Arguments

:: Int
$sel:httpStatus:DescribeJobRunResponse'

-> Text
$sel:jobName:DescribeJobRunResponse'

-> DescribeJobRunResponse

Create a value of DescribeJobRunResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:completedOn:DescribeJobRunResponse', describeJobRunResponse_completedOn - The date and time when the job completed processing.

$sel:state:DescribeJobRunResponse', describeJobRunResponse_state - The current state of the job run entity itself.

$sel:dataCatalogOutputs:DescribeJobRunResponse', describeJobRunResponse_dataCatalogOutputs - One or more artifacts that represent the Glue Data Catalog output from running the job.

$sel:startedOn:DescribeJobRunResponse', describeJobRunResponse_startedOn - The date and time when the job run began.

$sel:recipeReference:DescribeJobRunResponse', describeJobRunResponse_recipeReference - Undocumented member.

$sel:profileConfiguration:DescribeJobRunResponse', describeJobRunResponse_profileConfiguration - Configuration for profile jobs. Used to select columns, do evaluations, and override default parameters of evaluations. When configuration is null, the profile job will run with default settings.

$sel:databaseOutputs:DescribeJobRunResponse', describeJobRunResponse_databaseOutputs - Represents a list of JDBC database output objects which defines the output destination for a DataBrew recipe job to write into.

$sel:logGroupName:DescribeJobRunResponse', describeJobRunResponse_logGroupName - The name of an Amazon CloudWatch log group, where the job writes diagnostic messages when it runs.

$sel:outputs:DescribeJobRunResponse', describeJobRunResponse_outputs - One or more output artifacts from a job run.

$sel:runId:DescribeJobRun', describeJobRunResponse_runId - The unique identifier of the job run.

$sel:executionTime:DescribeJobRunResponse', describeJobRunResponse_executionTime - The amount of time, in seconds, during which the job run consumed resources.

$sel:logSubscription:DescribeJobRunResponse', describeJobRunResponse_logSubscription - The current status of Amazon CloudWatch logging for the job run.

$sel:startedBy:DescribeJobRunResponse', describeJobRunResponse_startedBy - The Amazon Resource Name (ARN) of the user who started the job run.

$sel:datasetName:DescribeJobRunResponse', describeJobRunResponse_datasetName - The name of the dataset for the job to process.

$sel:attempt:DescribeJobRunResponse', describeJobRunResponse_attempt - The number of times that DataBrew has attempted to run the job.

$sel:errorMessage:DescribeJobRunResponse', describeJobRunResponse_errorMessage - A message indicating an error (if any) that was encountered when the job ran.

$sel:jobSample:DescribeJobRunResponse', describeJobRunResponse_jobSample - Sample configuration for profile jobs only. Determines the number of rows on which the profile job will be executed. If a JobSample value is not provided, the default value will be used. The default value is CUSTOM_ROWS for the mode parameter and 20000 for the size parameter.

$sel:httpStatus:DescribeJobRunResponse', describeJobRunResponse_httpStatus - The response's http status code.

$sel:jobName:DescribeJobRunResponse', describeJobRunResponse_jobName - The name of the job being processed during this run.

Response Lenses

describeJobRunResponse_completedOn :: Lens' DescribeJobRunResponse (Maybe UTCTime)

The date and time when the job completed processing.

describeJobRunResponse_state :: Lens' DescribeJobRunResponse (Maybe JobRunState)

The current state of the job run entity itself.

describeJobRunResponse_dataCatalogOutputs :: Lens' DescribeJobRunResponse (Maybe (NonEmpty DataCatalogOutput))

One or more artifacts that represent the Glue Data Catalog output from running the job.

describeJobRunResponse_startedOn :: Lens' DescribeJobRunResponse (Maybe UTCTime)

The date and time when the job run began.

describeJobRunResponse_recipeReference :: Lens' DescribeJobRunResponse (Maybe RecipeReference)

Undocumented member.

describeJobRunResponse_profileConfiguration :: Lens' DescribeJobRunResponse (Maybe ProfileConfiguration)

Configuration for profile jobs. Used to select columns, do evaluations, and override default parameters of evaluations. When configuration is null, the profile job will run with default settings.

describeJobRunResponse_databaseOutputs :: Lens' DescribeJobRunResponse (Maybe (NonEmpty DatabaseOutput))

Represents a list of JDBC database output objects which defines the output destination for a DataBrew recipe job to write into.

describeJobRunResponse_logGroupName :: Lens' DescribeJobRunResponse (Maybe Text)

The name of an Amazon CloudWatch log group, where the job writes diagnostic messages when it runs.

describeJobRunResponse_outputs :: Lens' DescribeJobRunResponse (Maybe (NonEmpty Output))

One or more output artifacts from a job run.

describeJobRunResponse_runId :: Lens' DescribeJobRunResponse (Maybe Text)

The unique identifier of the job run.

describeJobRunResponse_executionTime :: Lens' DescribeJobRunResponse (Maybe Int)

The amount of time, in seconds, during which the job run consumed resources.

describeJobRunResponse_logSubscription :: Lens' DescribeJobRunResponse (Maybe LogSubscription)

The current status of Amazon CloudWatch logging for the job run.

describeJobRunResponse_startedBy :: Lens' DescribeJobRunResponse (Maybe Text)

The Amazon Resource Name (ARN) of the user who started the job run.

describeJobRunResponse_datasetName :: Lens' DescribeJobRunResponse (Maybe Text)

The name of the dataset for the job to process.

describeJobRunResponse_attempt :: Lens' DescribeJobRunResponse (Maybe Int)

The number of times that DataBrew has attempted to run the job.

describeJobRunResponse_errorMessage :: Lens' DescribeJobRunResponse (Maybe Text)

A message indicating an error (if any) that was encountered when the job ran.

describeJobRunResponse_jobSample :: Lens' DescribeJobRunResponse (Maybe JobSample)

Sample configuration for profile jobs only. Determines the number of rows on which the profile job will be executed. If a JobSample value is not provided, the default value will be used. The default value is CUSTOM_ROWS for the mode parameter and 20000 for the size parameter.

describeJobRunResponse_httpStatus :: Lens' DescribeJobRunResponse Int

The response's http status code.

describeJobRunResponse_jobName :: Lens' DescribeJobRunResponse Text

The name of the job being processed during this run.