Amazonka.DataBrew.ListJobs

Description

Lists all of the DataBrew jobs that are defined.

This operation returns paginated results.

Creating a Request

data ListJobs

See: newListJobs smart constructor.

Constructors

ListJobs'

Fields

nextToken :: Maybe Text

A token generated by DataBrew that specifies where to continue pagination if a previous request was truncated. To get the next set of pages, pass in the NextToken value from the response object of the previous page call.

projectName :: Maybe Text

The name of a project. Using this parameter indicates to return only those jobs that are associated with the specified project.

datasetName :: Maybe Text

The name of a dataset. Using this parameter indicates to return only those jobs that act on the specified dataset.

maxResults :: Maybe Natural

The maximum number of results to return in this request.

newListJobs :: ListJobs

Create a value of ListJobs with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:nextToken:ListJobs', listJobs_nextToken - A token generated by DataBrew that specifies where to continue pagination if a previous request was truncated. To get the next set of pages, pass in the NextToken value from the response object of the previous page call.

$sel:projectName:ListJobs', listJobs_projectName - The name of a project. Using this parameter indicates to return only those jobs that are associated with the specified project.

$sel:datasetName:ListJobs', listJobs_datasetName - The name of a dataset. Using this parameter indicates to return only those jobs that act on the specified dataset.

$sel:maxResults:ListJobs', listJobs_maxResults - The maximum number of results to return in this request.

Request Lenses

listJobs_nextToken :: Lens' ListJobs (Maybe Text)

A token generated by DataBrew that specifies where to continue pagination if a previous request was truncated. To get the next set of pages, pass in the NextToken value from the response object of the previous page call.

listJobs_projectName :: Lens' ListJobs (Maybe Text)

The name of a project. Using this parameter indicates to return only those jobs that are associated with the specified project.

listJobs_datasetName :: Lens' ListJobs (Maybe Text)

The name of a dataset. Using this parameter indicates to return only those jobs that act on the specified dataset.

listJobs_maxResults :: Lens' ListJobs (Maybe Natural)

The maximum number of results to return in this request.

Destructuring the Response

data ListJobsResponse

See: newListJobsResponse smart constructor.

Constructors

ListJobsResponse'

Fields

nextToken :: Maybe Text

A token that you can use in a subsequent call to retrieve the next set of results.

httpStatus :: Int

The response's http status code.

jobs :: [Job]

A list of jobs that are defined.

newListJobsResponse

Arguments

:: Int
$sel:httpStatus:ListJobsResponse'

-> ListJobsResponse

Create a value of ListJobsResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:nextToken:ListJobs', listJobsResponse_nextToken - A token that you can use in a subsequent call to retrieve the next set of results.

$sel:httpStatus:ListJobsResponse', listJobsResponse_httpStatus - The response's http status code.

$sel:jobs:ListJobsResponse', listJobsResponse_jobs - A list of jobs that are defined.

Response Lenses

listJobsResponse_nextToken :: Lens' ListJobsResponse (Maybe Text)

A token that you can use in a subsequent call to retrieve the next set of results.

listJobsResponse_httpStatus :: Lens' ListJobsResponse Int

The response's http status code.

listJobsResponse_jobs :: Lens' ListJobsResponse [Job]

A list of jobs that are defined.