Amazonka.DataBrew.ListRecipeVersions

Description

Lists the versions of a particular DataBrew recipe, except for LATEST_WORKING.

This operation returns paginated results.

Creating a Request

data ListRecipeVersions

See: newListRecipeVersions smart constructor.

Constructors

ListRecipeVersions'

Fields

nextToken :: Maybe Text

The token returned by a previous call to retrieve the next set of results.

maxResults :: Maybe Natural

The maximum number of results to return in this request.

name :: Text

The name of the recipe for which to return version information.

newListRecipeVersions

Arguments

:: Text
$sel:name:ListRecipeVersions'

-> ListRecipeVersions

Create a value of ListRecipeVersions with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:nextToken:ListRecipeVersions', listRecipeVersions_nextToken - The token returned by a previous call to retrieve the next set of results.

$sel:maxResults:ListRecipeVersions', listRecipeVersions_maxResults - The maximum number of results to return in this request.

$sel:name:ListRecipeVersions', listRecipeVersions_name - The name of the recipe for which to return version information.

Request Lenses

listRecipeVersions_nextToken :: Lens' ListRecipeVersions (Maybe Text)

The token returned by a previous call to retrieve the next set of results.

listRecipeVersions_maxResults :: Lens' ListRecipeVersions (Maybe Natural)

The maximum number of results to return in this request.

listRecipeVersions_name :: Lens' ListRecipeVersions Text

The name of the recipe for which to return version information.

Destructuring the Response

data ListRecipeVersionsResponse

See: newListRecipeVersionsResponse smart constructor.

Constructors

ListRecipeVersionsResponse'

Fields

nextToken :: Maybe Text

A token that you can use in a subsequent call to retrieve the next set of results.

httpStatus :: Int

The response's http status code.

recipes :: [Recipe]

A list of versions for the specified recipe.

newListRecipeVersionsResponse

Arguments

:: Int
$sel:httpStatus:ListRecipeVersionsResponse'

-> ListRecipeVersionsResponse

Create a value of ListRecipeVersionsResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:nextToken:ListRecipeVersions', listRecipeVersionsResponse_nextToken - A token that you can use in a subsequent call to retrieve the next set of results.

$sel:httpStatus:ListRecipeVersionsResponse', listRecipeVersionsResponse_httpStatus - The response's http status code.

$sel:recipes:ListRecipeVersionsResponse', listRecipeVersionsResponse_recipes - A list of versions for the specified recipe.

Response Lenses

listRecipeVersionsResponse_nextToken :: Lens' ListRecipeVersionsResponse (Maybe Text)

A token that you can use in a subsequent call to retrieve the next set of results.

listRecipeVersionsResponse_httpStatus :: Lens' ListRecipeVersionsResponse Int

The response's http status code.

listRecipeVersionsResponse_recipes :: Lens' ListRecipeVersionsResponse [Recipe]

A list of versions for the specified recipe.