Amazonka.DataBrew.SendProjectSessionAction

Description

Performs a recipe step within an interactive DataBrew session that's currently open.

Creating a Request

data SendProjectSessionAction

See: newSendProjectSessionAction smart constructor.

Constructors

SendProjectSessionAction'

Fields

stepIndex :: Maybe Natural

The index from which to preview a step. This index is used to preview the result of steps that have already been applied, so that the resulting view frame is from earlier in the view frame stack.

preview :: Maybe Bool

If true, the result of the recipe step will be returned, but not applied.

clientSessionId :: Maybe Text

A unique identifier for an interactive session that's currently open and ready for work. The action will be performed on this session.

recipeStep :: Maybe RecipeStep

viewFrame :: Maybe ViewFrame

name :: Text

The name of the project to apply the action to.

newSendProjectSessionAction

Arguments

:: Text
$sel:name:SendProjectSessionAction'

-> SendProjectSessionAction

Create a value of SendProjectSessionAction with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:stepIndex:SendProjectSessionAction', sendProjectSessionAction_stepIndex - The index from which to preview a step. This index is used to preview the result of steps that have already been applied, so that the resulting view frame is from earlier in the view frame stack.

$sel:preview:SendProjectSessionAction', sendProjectSessionAction_preview - If true, the result of the recipe step will be returned, but not applied.

$sel:clientSessionId:SendProjectSessionAction', sendProjectSessionAction_clientSessionId - A unique identifier for an interactive session that's currently open and ready for work. The action will be performed on this session.

$sel:recipeStep:SendProjectSessionAction', sendProjectSessionAction_recipeStep - Undocumented member.

$sel:viewFrame:SendProjectSessionAction', sendProjectSessionAction_viewFrame - Undocumented member.

$sel:name:SendProjectSessionAction', sendProjectSessionAction_name - The name of the project to apply the action to.

Request Lenses

sendProjectSessionAction_stepIndex :: Lens' SendProjectSessionAction (Maybe Natural)

The index from which to preview a step. This index is used to preview the result of steps that have already been applied, so that the resulting view frame is from earlier in the view frame stack.

sendProjectSessionAction_preview :: Lens' SendProjectSessionAction (Maybe Bool)

If true, the result of the recipe step will be returned, but not applied.

sendProjectSessionAction_clientSessionId :: Lens' SendProjectSessionAction (Maybe Text)

A unique identifier for an interactive session that's currently open and ready for work. The action will be performed on this session.

sendProjectSessionAction_recipeStep :: Lens' SendProjectSessionAction (Maybe RecipeStep)

Undocumented member.

sendProjectSessionAction_viewFrame :: Lens' SendProjectSessionAction (Maybe ViewFrame)

Undocumented member.

sendProjectSessionAction_name :: Lens' SendProjectSessionAction Text

The name of the project to apply the action to.

Destructuring the Response

data SendProjectSessionActionResponse

See: newSendProjectSessionActionResponse smart constructor.

Constructors

SendProjectSessionActionResponse'

Fields

actionId :: Maybe Int

A unique identifier for the action that was performed.

result :: Maybe Text

A message indicating the result of performing the action.

httpStatus :: Int

The response's http status code.

name :: Text

The name of the project that was affected by the action.

newSendProjectSessionActionResponse

Arguments

:: Int
$sel:httpStatus:SendProjectSessionActionResponse'

-> Text
$sel:name:SendProjectSessionAction'

-> SendProjectSessionActionResponse

Create a value of SendProjectSessionActionResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:actionId:SendProjectSessionActionResponse', sendProjectSessionActionResponse_actionId - A unique identifier for the action that was performed.

$sel:result:SendProjectSessionActionResponse', sendProjectSessionActionResponse_result - A message indicating the result of performing the action.

$sel:httpStatus:SendProjectSessionActionResponse', sendProjectSessionActionResponse_httpStatus - The response's http status code.

$sel:name:SendProjectSessionAction', sendProjectSessionActionResponse_name - The name of the project that was affected by the action.

Response Lenses

sendProjectSessionActionResponse_actionId :: Lens' SendProjectSessionActionResponse (Maybe Int)

A unique identifier for the action that was performed.

sendProjectSessionActionResponse_result :: Lens' SendProjectSessionActionResponse (Maybe Text)

A message indicating the result of performing the action.

sendProjectSessionActionResponse_httpStatus :: Lens' SendProjectSessionActionResponse Int

The response's http status code.

sendProjectSessionActionResponse_name :: Lens' SendProjectSessionActionResponse Text

The name of the project that was affected by the action.