Amazonka.DataBrew.StopJobRun

Description

Stops a particular run of a job.

Creating a Request

data StopJobRun

See: newStopJobRun smart constructor.

Constructors

StopJobRun'

Fields

name :: Text

The name of the job to be stopped.

runId :: Text

The ID of the job run to be stopped.

newStopJobRun

Arguments

:: Text
$sel:name:StopJobRun'

-> Text
$sel:runId:StopJobRun'

-> StopJobRun

Create a value of StopJobRun with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:name:StopJobRun', stopJobRun_name - The name of the job to be stopped.

$sel:runId:StopJobRun', stopJobRun_runId - The ID of the job run to be stopped.

Request Lenses

stopJobRun_name :: Lens' StopJobRun Text

The name of the job to be stopped.

stopJobRun_runId :: Lens' StopJobRun Text

The ID of the job run to be stopped.

Destructuring the Response

data StopJobRunResponse

See: newStopJobRunResponse smart constructor.

Constructors

StopJobRunResponse'

Fields

httpStatus :: Int

The response's http status code.

runId :: Text

The ID of the job run that you stopped.

newStopJobRunResponse

Arguments

:: Int
$sel:httpStatus:StopJobRunResponse'

-> Text
$sel:runId:StopJobRun'

-> StopJobRunResponse

Create a value of StopJobRunResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:httpStatus:StopJobRunResponse', stopJobRunResponse_httpStatus - The response's http status code.

$sel:runId:StopJobRun', stopJobRunResponse_runId - The ID of the job run that you stopped.

Response Lenses

stopJobRunResponse_httpStatus :: Lens' StopJobRunResponse Int

The response's http status code.

stopJobRunResponse_runId :: Lens' StopJobRunResponse Text

The ID of the job run that you stopped.