Amazonka.DataBrew.Types.CsvOptions

Description

Documentation

data CsvOptions

Represents a set of options that define how DataBrew will read a comma-separated value (CSV) file when creating a dataset from that file.

See: newCsvOptions smart constructor.

Constructors

CsvOptions'

Fields

headerRow :: Maybe Bool

A variable that specifies whether the first row in the file is parsed as the header. If this value is false, column names are auto-generated.

delimiter :: Maybe Text

A single character that specifies the delimiter being used in the CSV file.

newCsvOptions :: CsvOptions

Create a value of CsvOptions with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:headerRow:CsvOptions', csvOptions_headerRow - A variable that specifies whether the first row in the file is parsed as the header. If this value is false, column names are auto-generated.

$sel:delimiter:CsvOptions', csvOptions_delimiter - A single character that specifies the delimiter being used in the CSV file.

csvOptions_headerRow :: Lens' CsvOptions (Maybe Bool)

A variable that specifies whether the first row in the file is parsed as the header. If this value is false, column names are auto-generated.

csvOptions_delimiter :: Lens' CsvOptions (Maybe Text)

A single character that specifies the delimiter being used in the CSV file.