Amazonka.DataBrew.Types.DatetimeOptions

Description

Documentation

data DatetimeOptions

Represents additional options for correct interpretation of datetime parameters used in the Amazon S3 path of a dataset.

See: newDatetimeOptions smart constructor.

Constructors

DatetimeOptions'

Fields

timezoneOffset :: Maybe Text

Optional value for a timezone offset of the datetime parameter value in the Amazon S3 path. Shouldn't be used if Format for this parameter includes timezone fields. If no offset specified, UTC is assumed.

localeCode :: Maybe Text

Optional value for a non-US locale code, needed for correct interpretation of some date formats.

format :: Text

Required option, that defines the datetime format used for a date parameter in the Amazon S3 path. Should use only supported datetime specifiers and separation characters, all literal a-z or A-Z characters should be escaped with single quotes. E.g. "MM.dd.yyyy-'at'-HH:mm".

newDatetimeOptions

Arguments

:: Text
$sel:format:DatetimeOptions'

-> DatetimeOptions

Create a value of DatetimeOptions with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:timezoneOffset:DatetimeOptions', datetimeOptions_timezoneOffset - Optional value for a timezone offset of the datetime parameter value in the Amazon S3 path. Shouldn't be used if Format for this parameter includes timezone fields. If no offset specified, UTC is assumed.

$sel:localeCode:DatetimeOptions', datetimeOptions_localeCode - Optional value for a non-US locale code, needed for correct interpretation of some date formats.

$sel:format:DatetimeOptions', datetimeOptions_format - Required option, that defines the datetime format used for a date parameter in the Amazon S3 path. Should use only supported datetime specifiers and separation characters, all literal a-z or A-Z characters should be escaped with single quotes. E.g. "MM.dd.yyyy-'at'-HH:mm".

datetimeOptions_timezoneOffset :: Lens' DatetimeOptions (Maybe Text)

Optional value for a timezone offset of the datetime parameter value in the Amazon S3 path. Shouldn't be used if Format for this parameter includes timezone fields. If no offset specified, UTC is assumed.

datetimeOptions_localeCode :: Lens' DatetimeOptions (Maybe Text)

Optional value for a non-US locale code, needed for correct interpretation of some date formats.

datetimeOptions_format :: Lens' DatetimeOptions Text

Required option, that defines the datetime format used for a date parameter in the Amazon S3 path. Should use only supported datetime specifiers and separation characters, all literal a-z or A-Z characters should be escaped with single quotes. E.g. "MM.dd.yyyy-'at'-HH:mm".