Amazonka.DataBrew.Types.FilesLimit

Description

Documentation

data FilesLimit

Represents a limit imposed on number of Amazon S3 files that should be selected for a dataset from a connected Amazon S3 path.

See: newFilesLimit smart constructor.

Constructors

FilesLimit'

Fields

orderedBy :: Maybe OrderedBy

A criteria to use for Amazon S3 files sorting before their selection. By default uses LAST_MODIFIED_DATE as a sorting criteria. Currently it's the only allowed value.

order :: Maybe Order

A criteria to use for Amazon S3 files sorting before their selection. By default uses DESCENDING order, i.e. most recent files are selected first. Anotherpossible value is ASCENDING.

maxFiles :: Natural

The number of Amazon S3 files to select.

newFilesLimit

Arguments

:: Natural
$sel:maxFiles:FilesLimit'

-> FilesLimit

Create a value of FilesLimit with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:orderedBy:FilesLimit', filesLimit_orderedBy - A criteria to use for Amazon S3 files sorting before their selection. By default uses LAST_MODIFIED_DATE as a sorting criteria. Currently it's the only allowed value.

$sel:order:FilesLimit', filesLimit_order - A criteria to use for Amazon S3 files sorting before their selection. By default uses DESCENDING order, i.e. most recent files are selected first. Anotherpossible value is ASCENDING.

$sel:maxFiles:FilesLimit', filesLimit_maxFiles - The number of Amazon S3 files to select.

filesLimit_orderedBy :: Lens' FilesLimit (Maybe OrderedBy)

A criteria to use for Amazon S3 files sorting before their selection. By default uses LAST_MODIFIED_DATE as a sorting criteria. Currently it's the only allowed value.

filesLimit_order :: Lens' FilesLimit (Maybe Order)

A criteria to use for Amazon S3 files sorting before their selection. By default uses DESCENDING order, i.e. most recent files are selected first. Anotherpossible value is ASCENDING.

filesLimit_maxFiles :: Lens' FilesLimit Natural

The number of Amazon S3 files to select.