Amazonka.DataBrew.Types.FilterExpression

Description

Documentation

data FilterExpression

Represents a structure for defining parameter conditions. Supported conditions are described here: Supported conditions for dynamic datasets in the Glue DataBrew Developer Guide.

See: newFilterExpression smart constructor.

Constructors

FilterExpression'

Fields

expression :: Text

The expression which includes condition names followed by substitution variables, possibly grouped and combined with other conditions. For example, "(starts_with :prefix1 or starts_with :prefix2) and (ends_with :suffix1 or ends_with :suffix2)". Substitution variables should start with ':' symbol.

valuesMap :: HashMap Text Text

The map of substitution variable names to their values used in this filter expression.

newFilterExpression

Arguments

:: Text
$sel:expression:FilterExpression'

-> FilterExpression

Create a value of FilterExpression with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:expression:FilterExpression', filterExpression_expression - The expression which includes condition names followed by substitution variables, possibly grouped and combined with other conditions. For example, "(starts_with :prefix1 or starts_with :prefix2) and (ends_with :suffix1 or ends_with :suffix2)". Substitution variables should start with ':' symbol.

$sel:valuesMap:FilterExpression', filterExpression_valuesMap - The map of substitution variable names to their values used in this filter expression.

filterExpression_expression :: Lens' FilterExpression Text

The expression which includes condition names followed by substitution variables, possibly grouped and combined with other conditions. For example, "(starts_with :prefix1 or starts_with :prefix2) and (ends_with :suffix1 or ends_with :suffix2)". Substitution variables should start with ':' symbol.

filterExpression_valuesMap :: Lens' FilterExpression (HashMap Text Text)

The map of substitution variable names to their values used in this filter expression.