Amazonka.DataBrew.Types.Job

Description

Documentation

data Job

Represents all of the attributes of a DataBrew job.

See: newJob smart constructor.

Constructors

Job'

Fields

dataCatalogOutputs :: Maybe (NonEmpty DataCatalogOutput)

One or more artifacts that represent the Glue Data Catalog output from running the job.

lastModifiedDate :: Maybe POSIX

The modification date and time of the job.

createDate :: Maybe POSIX

The date and time that the job was created.

recipeReference :: Maybe RecipeReference

A set of steps that the job runs.

createdBy :: Maybe Text

The Amazon Resource Name (ARN) of the user who created the job.

databaseOutputs :: Maybe (NonEmpty DatabaseOutput)

Represents a list of JDBC database output objects which defines the output destination for a DataBrew recipe job to write into.

accountId :: Maybe Text

The ID of the Amazon Web Services account that owns the job.

encryptionMode :: Maybe EncryptionMode

The encryption mode for the job, which can be one of the following:

SSE-KMS - Server-side encryption with keys managed by KMS.

SSE-S3 - Server-side encryption with keys managed by Amazon S3.

outputs :: Maybe (NonEmpty Output)

One or more artifacts that represent output from running the job.

resourceArn :: Maybe Text

The unique Amazon Resource Name (ARN) for the job.

logSubscription :: Maybe LogSubscription

The current status of Amazon CloudWatch logging for the job.

projectName :: Maybe Text

The name of the project that the job is associated with.

maxRetries :: Maybe Natural

The maximum number of times to retry the job after a job run fails.

datasetName :: Maybe Text

A dataset that the job is to process.

encryptionKeyArn :: Maybe Text

The Amazon Resource Name (ARN) of an encryption key that is used to protect the job output. For more information, see Encrypting data written by DataBrew jobs

maxCapacity :: Maybe Int

The maximum number of nodes that can be consumed when the job processes data.

lastModifiedBy :: Maybe Text

The Amazon Resource Name (ARN) of the user who last modified the job.

type' :: Maybe JobType

The job type of the job, which must be one of the following:

PROFILE - A job to analyze a dataset, to determine its size, data types, data distribution, and more.

RECIPE - A job to apply one or more transformations to a dataset.

timeout :: Maybe Natural

The job's timeout in minutes. A job that attempts to run longer than this timeout period ends with a status of TIMEOUT.

tags :: Maybe (HashMap Text Text)

Metadata tags that have been applied to the job.

roleArn :: Maybe Text

The Amazon Resource Name (ARN) of the role to be assumed for this job.

jobSample :: Maybe JobSample

A sample configuration for profile jobs only, which determines the number of rows on which the profile job is run. If a JobSample value isn't provided, the default value is used. The default value is CUSTOM_ROWS for the mode parameter and 20,000 for the size parameter.

name :: Text

The unique name of the job.

newJob

Arguments

:: Text
$sel:name:Job'

-> Job

Create a value of Job with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:dataCatalogOutputs:Job', job_dataCatalogOutputs - One or more artifacts that represent the Glue Data Catalog output from running the job.

$sel:lastModifiedDate:Job', job_lastModifiedDate - The modification date and time of the job.

$sel:createDate:Job', job_createDate - The date and time that the job was created.

$sel:recipeReference:Job', job_recipeReference - A set of steps that the job runs.

$sel:createdBy:Job', job_createdBy - The Amazon Resource Name (ARN) of the user who created the job.

$sel:databaseOutputs:Job', job_databaseOutputs - Represents a list of JDBC database output objects which defines the output destination for a DataBrew recipe job to write into.

$sel:accountId:Job', job_accountId - The ID of the Amazon Web Services account that owns the job.

$sel:encryptionMode:Job', job_encryptionMode - The encryption mode for the job, which can be one of the following:

SSE-KMS - Server-side encryption with keys managed by KMS.

SSE-S3 - Server-side encryption with keys managed by Amazon S3.

$sel:outputs:Job', job_outputs - One or more artifacts that represent output from running the job.

$sel:resourceArn:Job', job_resourceArn - The unique Amazon Resource Name (ARN) for the job.

$sel:logSubscription:Job', job_logSubscription - The current status of Amazon CloudWatch logging for the job.

$sel:projectName:Job', job_projectName - The name of the project that the job is associated with.

$sel:maxRetries:Job', job_maxRetries - The maximum number of times to retry the job after a job run fails.

$sel:datasetName:Job', job_datasetName - A dataset that the job is to process.

$sel:encryptionKeyArn:Job', job_encryptionKeyArn - The Amazon Resource Name (ARN) of an encryption key that is used to protect the job output. For more information, see Encrypting data written by DataBrew jobs

$sel:maxCapacity:Job', job_maxCapacity - The maximum number of nodes that can be consumed when the job processes data.

$sel:lastModifiedBy:Job', job_lastModifiedBy - The Amazon Resource Name (ARN) of the user who last modified the job.

$sel:type':Job', job_type - The job type of the job, which must be one of the following:

PROFILE - A job to analyze a dataset, to determine its size, data types, data distribution, and more.

RECIPE - A job to apply one or more transformations to a dataset.

$sel:timeout:Job', job_timeout - The job's timeout in minutes. A job that attempts to run longer than this timeout period ends with a status of TIMEOUT.

$sel:tags:Job', job_tags - Metadata tags that have been applied to the job.

$sel:roleArn:Job', job_roleArn - The Amazon Resource Name (ARN) of the role to be assumed for this job.

$sel:jobSample:Job', job_jobSample - A sample configuration for profile jobs only, which determines the number of rows on which the profile job is run. If a JobSample value isn't provided, the default value is used. The default value is CUSTOM_ROWS for the mode parameter and 20,000 for the size parameter.

$sel:name:Job', job_name - The unique name of the job.

job_dataCatalogOutputs :: Lens' Job (Maybe (NonEmpty DataCatalogOutput))

One or more artifacts that represent the Glue Data Catalog output from running the job.

job_lastModifiedDate :: Lens' Job (Maybe UTCTime)

The modification date and time of the job.

job_createDate :: Lens' Job (Maybe UTCTime)

The date and time that the job was created.

job_recipeReference :: Lens' Job (Maybe RecipeReference)

A set of steps that the job runs.

job_createdBy :: Lens' Job (Maybe Text)

The Amazon Resource Name (ARN) of the user who created the job.

job_databaseOutputs :: Lens' Job (Maybe (NonEmpty DatabaseOutput))

Represents a list of JDBC database output objects which defines the output destination for a DataBrew recipe job to write into.

job_accountId :: Lens' Job (Maybe Text)

The ID of the Amazon Web Services account that owns the job.

job_encryptionMode :: Lens' Job (Maybe EncryptionMode)

The encryption mode for the job, which can be one of the following:

SSE-KMS - Server-side encryption with keys managed by KMS.

SSE-S3 - Server-side encryption with keys managed by Amazon S3.

job_outputs :: Lens' Job (Maybe (NonEmpty Output))

One or more artifacts that represent output from running the job.

job_resourceArn :: Lens' Job (Maybe Text)

The unique Amazon Resource Name (ARN) for the job.

job_logSubscription :: Lens' Job (Maybe LogSubscription)

The current status of Amazon CloudWatch logging for the job.

job_projectName :: Lens' Job (Maybe Text)

The name of the project that the job is associated with.

job_maxRetries :: Lens' Job (Maybe Natural)

The maximum number of times to retry the job after a job run fails.

job_datasetName :: Lens' Job (Maybe Text)

A dataset that the job is to process.

job_encryptionKeyArn :: Lens' Job (Maybe Text)

The Amazon Resource Name (ARN) of an encryption key that is used to protect the job output. For more information, see Encrypting data written by DataBrew jobs

job_maxCapacity :: Lens' Job (Maybe Int)

The maximum number of nodes that can be consumed when the job processes data.

job_lastModifiedBy :: Lens' Job (Maybe Text)

The Amazon Resource Name (ARN) of the user who last modified the job.

job_type :: Lens' Job (Maybe JobType)

The job type of the job, which must be one of the following:

PROFILE - A job to analyze a dataset, to determine its size, data types, data distribution, and more.

RECIPE - A job to apply one or more transformations to a dataset.

job_timeout :: Lens' Job (Maybe Natural)

The job's timeout in minutes. A job that attempts to run longer than this timeout period ends with a status of TIMEOUT.

job_tags :: Lens' Job (Maybe (HashMap Text Text))

Metadata tags that have been applied to the job.

job_roleArn :: Lens' Job (Maybe Text)

The Amazon Resource Name (ARN) of the role to be assumed for this job.

job_jobSample :: Lens' Job (Maybe JobSample)

A sample configuration for profile jobs only, which determines the number of rows on which the profile job is run. If a JobSample value isn't provided, the default value is used. The default value is CUSTOM_ROWS for the mode parameter and 20,000 for the size parameter.

job_name :: Lens' Job Text

The unique name of the job.