Amazonka.DataBrew.Types.JobSample

Description

Documentation

data JobSample

A sample configuration for profile jobs only, which determines the number of rows on which the profile job is run. If a JobSample value isn't provided, the default is used. The default value is CUSTOM_ROWS for the mode parameter and 20,000 for the size parameter.

See: newJobSample smart constructor.

Constructors

JobSample'

Fields

size :: Maybe Integer

The Size parameter is only required when the mode is CUSTOM_ROWS. The profile job is run on the specified number of rows. The maximum value for size is Long.MAX_VALUE.

Long.MAX_VALUE = 9223372036854775807

mode :: Maybe SampleMode

A value that determines whether the profile job is run on the entire dataset or a specified number of rows. This value must be one of the following:

FULL_DATASET - The profile job is run on the entire dataset.

CUSTOM_ROWS - The profile job is run on the number of rows specified in the Size parameter.

newJobSample :: JobSample

Create a value of JobSample with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:size:JobSample', jobSample_size - The Size parameter is only required when the mode is CUSTOM_ROWS. The profile job is run on the specified number of rows. The maximum value for size is Long.MAX_VALUE.

Long.MAX_VALUE = 9223372036854775807

$sel:mode:JobSample', jobSample_mode - A value that determines whether the profile job is run on the entire dataset or a specified number of rows. This value must be one of the following:

FULL_DATASET - The profile job is run on the entire dataset.

CUSTOM_ROWS - The profile job is run on the number of rows specified in the Size parameter.

jobSample_size :: Lens' JobSample (Maybe Integer)

The Size parameter is only required when the mode is CUSTOM_ROWS. The profile job is run on the specified number of rows. The maximum value for size is Long.MAX_VALUE.

Long.MAX_VALUE = 9223372036854775807

jobSample_mode :: Lens' JobSample (Maybe SampleMode)

A value that determines whether the profile job is run on the entire dataset or a specified number of rows. This value must be one of the following:

FULL_DATASET - The profile job is run on the entire dataset.

CUSTOM_ROWS - The profile job is run on the number of rows specified in the Size parameter.