Amazonka.DataBrew.Types.Output

Description

Documentation

data Output

Represents options that specify how and where in Amazon S3 DataBrew writes the output generated by recipe jobs or profile jobs.

See: newOutput smart constructor.

Constructors

Output'

Fields

partitionColumns :: Maybe [Text]

The names of one or more partition columns for the output of the job.

formatOptions :: Maybe OutputFormatOptions

Represents options that define how DataBrew formats job output files.

format :: Maybe OutputFormat

The data format of the output of the job.

compressionFormat :: Maybe CompressionFormat

The compression algorithm used to compress the output text of the job.

overwrite :: Maybe Bool

A value that, if true, means that any data in the location specified for output is overwritten with new output.

location :: S3Location

The location in Amazon S3 where the job writes its output.

newOutput

Arguments

:: S3Location
$sel:location:Output'

-> Output

Create a value of Output with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:partitionColumns:Output', output_partitionColumns - The names of one or more partition columns for the output of the job.

$sel:formatOptions:Output', output_formatOptions - Represents options that define how DataBrew formats job output files.

$sel:format:Output', output_format - The data format of the output of the job.

$sel:compressionFormat:Output', output_compressionFormat - The compression algorithm used to compress the output text of the job.

$sel:overwrite:Output', output_overwrite - A value that, if true, means that any data in the location specified for output is overwritten with new output.

$sel:location:Output', output_location - The location in Amazon S3 where the job writes its output.

output_partitionColumns :: Lens' Output (Maybe [Text])

The names of one or more partition columns for the output of the job.

output_formatOptions :: Lens' Output (Maybe OutputFormatOptions)

Represents options that define how DataBrew formats job output files.

output_format :: Lens' Output (Maybe OutputFormat)

The data format of the output of the job.

output_compressionFormat :: Lens' Output (Maybe CompressionFormat)

The compression algorithm used to compress the output text of the job.

output_overwrite :: Lens' Output (Maybe Bool)

A value that, if true, means that any data in the location specified for output is overwritten with new output.

output_location :: Lens' Output S3Location

The location in Amazon S3 where the job writes its output.