Amazonka.DataBrew.Types.Project

Description

Documentation

data Project

Represents all of the attributes of a DataBrew project.

See: newProject smart constructor.

Constructors

Project'

Fields

lastModifiedDate :: Maybe POSIX

The last modification date and time for the project.

openDate :: Maybe POSIX

The date and time when the project was opened.

createDate :: Maybe POSIX

The date and time that the project was created.

createdBy :: Maybe Text

The Amazon Resource Name (ARN) of the user who crated the project.

accountId :: Maybe Text

The ID of the Amazon Web Services account that owns the project.

resourceArn :: Maybe Text

The Amazon Resource Name (ARN) for the project.

datasetName :: Maybe Text

The dataset that the project is to act upon.

lastModifiedBy :: Maybe Text

The Amazon Resource Name (ARN) of the user who last modified the project.

sample :: Maybe Sample

The sample size and sampling type to apply to the data. If this parameter isn't specified, then the sample consists of the first 500 rows from the dataset.

openedBy :: Maybe Text

The Amazon Resource Name (ARN) of the user that opened the project for use.

tags :: Maybe (HashMap Text Text)

Metadata tags that have been applied to the project.

roleArn :: Maybe Text

The Amazon Resource Name (ARN) of the role that will be assumed for this project.

name :: Text

The unique name of a project.

recipeName :: Text

The name of a recipe that will be developed during a project session.

newProject

Arguments

:: Text
$sel:name:Project'

-> Text
$sel:recipeName:Project'

-> Project

Create a value of Project with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:lastModifiedDate:Project', project_lastModifiedDate - The last modification date and time for the project.

$sel:openDate:Project', project_openDate - The date and time when the project was opened.

$sel:createDate:Project', project_createDate - The date and time that the project was created.

$sel:createdBy:Project', project_createdBy - The Amazon Resource Name (ARN) of the user who crated the project.

$sel:accountId:Project', project_accountId - The ID of the Amazon Web Services account that owns the project.

$sel:resourceArn:Project', project_resourceArn - The Amazon Resource Name (ARN) for the project.

$sel:datasetName:Project', project_datasetName - The dataset that the project is to act upon.

$sel:lastModifiedBy:Project', project_lastModifiedBy - The Amazon Resource Name (ARN) of the user who last modified the project.

$sel:sample:Project', project_sample - The sample size and sampling type to apply to the data. If this parameter isn't specified, then the sample consists of the first 500 rows from the dataset.

$sel:openedBy:Project', project_openedBy - The Amazon Resource Name (ARN) of the user that opened the project for use.

$sel:tags:Project', project_tags - Metadata tags that have been applied to the project.

$sel:roleArn:Project', project_roleArn - The Amazon Resource Name (ARN) of the role that will be assumed for this project.

$sel:name:Project', project_name - The unique name of a project.

$sel:recipeName:Project', project_recipeName - The name of a recipe that will be developed during a project session.

project_lastModifiedDate :: Lens' Project (Maybe UTCTime)

The last modification date and time for the project.

project_openDate :: Lens' Project (Maybe UTCTime)

The date and time when the project was opened.

project_createDate :: Lens' Project (Maybe UTCTime)

The date and time that the project was created.

project_createdBy :: Lens' Project (Maybe Text)

The Amazon Resource Name (ARN) of the user who crated the project.

project_accountId :: Lens' Project (Maybe Text)

The ID of the Amazon Web Services account that owns the project.

project_resourceArn :: Lens' Project (Maybe Text)

The Amazon Resource Name (ARN) for the project.

project_datasetName :: Lens' Project (Maybe Text)

The dataset that the project is to act upon.

project_lastModifiedBy :: Lens' Project (Maybe Text)

The Amazon Resource Name (ARN) of the user who last modified the project.

project_sample :: Lens' Project (Maybe Sample)

The sample size and sampling type to apply to the data. If this parameter isn't specified, then the sample consists of the first 500 rows from the dataset.

project_openedBy :: Lens' Project (Maybe Text)

The Amazon Resource Name (ARN) of the user that opened the project for use.

project_tags :: Lens' Project (Maybe (HashMap Text Text))

Metadata tags that have been applied to the project.

project_roleArn :: Lens' Project (Maybe Text)

The Amazon Resource Name (ARN) of the role that will be assumed for this project.

project_name :: Lens' Project Text

The unique name of a project.

project_recipeName :: Lens' Project Text

The name of a recipe that will be developed during a project session.