Amazonka.DataBrew.Types.Recipe

Description

Documentation

data Recipe

Represents one or more actions to be performed on a DataBrew dataset.

See: newRecipe smart constructor.

Constructors

Recipe'

Fields

lastModifiedDate :: Maybe POSIX

The last modification date and time of the recipe.

createDate :: Maybe POSIX

The date and time that the recipe was created.

publishedBy :: Maybe Text

The Amazon Resource Name (ARN) of the user who published the recipe.

createdBy :: Maybe Text

The Amazon Resource Name (ARN) of the user who created the recipe.

steps :: Maybe [RecipeStep]

A list of steps that are defined by the recipe.

publishedDate :: Maybe POSIX

The date and time when the recipe was published.

resourceArn :: Maybe Text

The Amazon Resource Name (ARN) for the recipe.

recipeVersion :: Maybe Text

The identifier for the version for the recipe. Must be one of the following:

Numeric version (X.Y) - X and Y stand for major and minor version numbers. The maximum length of each is 6 digits, and neither can be negative values. Both X and Y are required, and "0.0" isn't a valid version.

LATEST_WORKING - the most recent valid version being developed in a DataBrew project.

LATEST_PUBLISHED - the most recent published version.

projectName :: Maybe Text

The name of the project that the recipe is associated with.

lastModifiedBy :: Maybe Text

The Amazon Resource Name (ARN) of the user who last modified the recipe.

description :: Maybe Text

The description of the recipe.

tags :: Maybe (HashMap Text Text)

Metadata tags that have been applied to the recipe.

name :: Text

The unique name for the recipe.

newRecipe

Arguments

:: Text
$sel:name:Recipe'

-> Recipe

Create a value of Recipe with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:lastModifiedDate:Recipe', recipe_lastModifiedDate - The last modification date and time of the recipe.

$sel:createDate:Recipe', recipe_createDate - The date and time that the recipe was created.

$sel:publishedBy:Recipe', recipe_publishedBy - The Amazon Resource Name (ARN) of the user who published the recipe.

$sel:createdBy:Recipe', recipe_createdBy - The Amazon Resource Name (ARN) of the user who created the recipe.

$sel:steps:Recipe', recipe_steps - A list of steps that are defined by the recipe.

$sel:publishedDate:Recipe', recipe_publishedDate - The date and time when the recipe was published.

$sel:resourceArn:Recipe', recipe_resourceArn - The Amazon Resource Name (ARN) for the recipe.

$sel:recipeVersion:Recipe', recipe_recipeVersion - The identifier for the version for the recipe. Must be one of the following:

Numeric version (X.Y) - X and Y stand for major and minor version numbers. The maximum length of each is 6 digits, and neither can be negative values. Both X and Y are required, and "0.0" isn't a valid version.

LATEST_WORKING - the most recent valid version being developed in a DataBrew project.

LATEST_PUBLISHED - the most recent published version.

$sel:projectName:Recipe', recipe_projectName - The name of the project that the recipe is associated with.

$sel:lastModifiedBy:Recipe', recipe_lastModifiedBy - The Amazon Resource Name (ARN) of the user who last modified the recipe.

$sel:description:Recipe', recipe_description - The description of the recipe.

$sel:tags:Recipe', recipe_tags - Metadata tags that have been applied to the recipe.

$sel:name:Recipe', recipe_name - The unique name for the recipe.

recipe_lastModifiedDate :: Lens' Recipe (Maybe UTCTime)

The last modification date and time of the recipe.

recipe_createDate :: Lens' Recipe (Maybe UTCTime)

The date and time that the recipe was created.

recipe_publishedBy :: Lens' Recipe (Maybe Text)

The Amazon Resource Name (ARN) of the user who published the recipe.

recipe_createdBy :: Lens' Recipe (Maybe Text)

The Amazon Resource Name (ARN) of the user who created the recipe.

recipe_steps :: Lens' Recipe (Maybe [RecipeStep])

A list of steps that are defined by the recipe.

recipe_publishedDate :: Lens' Recipe (Maybe UTCTime)

The date and time when the recipe was published.

recipe_resourceArn :: Lens' Recipe (Maybe Text)

The Amazon Resource Name (ARN) for the recipe.

recipe_recipeVersion :: Lens' Recipe (Maybe Text)

The identifier for the version for the recipe. Must be one of the following:

Numeric version (X.Y) - X and Y stand for major and minor version numbers. The maximum length of each is 6 digits, and neither can be negative values. Both X and Y are required, and "0.0" isn't a valid version.

LATEST_WORKING - the most recent valid version being developed in a DataBrew project.

LATEST_PUBLISHED - the most recent published version.

recipe_projectName :: Lens' Recipe (Maybe Text)

The name of the project that the recipe is associated with.

recipe_lastModifiedBy :: Lens' Recipe (Maybe Text)

The Amazon Resource Name (ARN) of the user who last modified the recipe.

recipe_description :: Lens' Recipe (Maybe Text)

The description of the recipe.

recipe_tags :: Lens' Recipe (Maybe (HashMap Text Text))

Metadata tags that have been applied to the recipe.

recipe_name :: Lens' Recipe Text

The unique name for the recipe.