Amazonka.DataBrew.Types.Schedule

Description

Documentation

data Schedule

Represents one or more dates and times when a job is to run.

See: newSchedule smart constructor.

Constructors

Schedule'

Fields

lastModifiedDate :: Maybe POSIX

The date and time when the schedule was last modified.

createDate :: Maybe POSIX

The date and time that the schedule was created.

createdBy :: Maybe Text

The Amazon Resource Name (ARN) of the user who created the schedule.

accountId :: Maybe Text

The ID of the Amazon Web Services account that owns the schedule.

resourceArn :: Maybe Text

The Amazon Resource Name (ARN) of the schedule.

cronExpression :: Maybe Text

The dates and times when the job is to run. For more information, see Cron expressions in the Glue DataBrew Developer Guide.

lastModifiedBy :: Maybe Text

The Amazon Resource Name (ARN) of the user who last modified the schedule.

jobNames :: Maybe [Text]

A list of jobs to be run, according to the schedule.

tags :: Maybe (HashMap Text Text)

Metadata tags that have been applied to the schedule.

name :: Text

The name of the schedule.

newSchedule

Arguments

:: Text
$sel:name:Schedule'

-> Schedule

Create a value of Schedule with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:lastModifiedDate:Schedule', schedule_lastModifiedDate - The date and time when the schedule was last modified.

$sel:createDate:Schedule', schedule_createDate - The date and time that the schedule was created.

$sel:createdBy:Schedule', schedule_createdBy - The Amazon Resource Name (ARN) of the user who created the schedule.

$sel:accountId:Schedule', schedule_accountId - The ID of the Amazon Web Services account that owns the schedule.

$sel:resourceArn:Schedule', schedule_resourceArn - The Amazon Resource Name (ARN) of the schedule.

$sel:cronExpression:Schedule', schedule_cronExpression - The dates and times when the job is to run. For more information, see Cron expressions in the Glue DataBrew Developer Guide.

$sel:lastModifiedBy:Schedule', schedule_lastModifiedBy - The Amazon Resource Name (ARN) of the user who last modified the schedule.

$sel:jobNames:Schedule', schedule_jobNames - A list of jobs to be run, according to the schedule.

$sel:tags:Schedule', schedule_tags - Metadata tags that have been applied to the schedule.

$sel:name:Schedule', schedule_name - The name of the schedule.

schedule_lastModifiedDate :: Lens' Schedule (Maybe UTCTime)

The date and time when the schedule was last modified.

schedule_createDate :: Lens' Schedule (Maybe UTCTime)

The date and time that the schedule was created.

schedule_createdBy :: Lens' Schedule (Maybe Text)

The Amazon Resource Name (ARN) of the user who created the schedule.

schedule_accountId :: Lens' Schedule (Maybe Text)

The ID of the Amazon Web Services account that owns the schedule.

schedule_resourceArn :: Lens' Schedule (Maybe Text)

The Amazon Resource Name (ARN) of the schedule.

schedule_cronExpression :: Lens' Schedule (Maybe Text)

The dates and times when the job is to run. For more information, see Cron expressions in the Glue DataBrew Developer Guide.

schedule_lastModifiedBy :: Lens' Schedule (Maybe Text)

The Amazon Resource Name (ARN) of the user who last modified the schedule.

schedule_jobNames :: Lens' Schedule (Maybe [Text])

A list of jobs to be run, according to the schedule.

schedule_tags :: Lens' Schedule (Maybe (HashMap Text Text))

Metadata tags that have been applied to the schedule.

schedule_name :: Lens' Schedule Text

The name of the schedule.