Amazonka.DataBrew.UpdateDataset

Description

Modifies the definition of an existing DataBrew dataset.

Creating a Request

data UpdateDataset

See: newUpdateDataset smart constructor.

Constructors

UpdateDataset'

Fields

pathOptions :: Maybe PathOptions

A set of options that defines how DataBrew interprets an Amazon S3 path of the dataset.

formatOptions :: Maybe FormatOptions

format :: Maybe InputFormat

The file format of a dataset that is created from an Amazon S3 file or folder.

name :: Text

The name of the dataset to be updated.

input :: Input

newUpdateDataset

Arguments

:: Text
$sel:name:UpdateDataset'

-> Input
$sel:input:UpdateDataset'

-> UpdateDataset

Create a value of UpdateDataset with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:pathOptions:UpdateDataset', updateDataset_pathOptions - A set of options that defines how DataBrew interprets an Amazon S3 path of the dataset.

$sel:formatOptions:UpdateDataset', updateDataset_formatOptions - Undocumented member.

$sel:format:UpdateDataset', updateDataset_format - The file format of a dataset that is created from an Amazon S3 file or folder.

$sel:name:UpdateDataset', updateDataset_name - The name of the dataset to be updated.

$sel:input:UpdateDataset', updateDataset_input - Undocumented member.

Request Lenses

updateDataset_pathOptions :: Lens' UpdateDataset (Maybe PathOptions)

A set of options that defines how DataBrew interprets an Amazon S3 path of the dataset.

updateDataset_formatOptions :: Lens' UpdateDataset (Maybe FormatOptions)

Undocumented member.

updateDataset_format :: Lens' UpdateDataset (Maybe InputFormat)

The file format of a dataset that is created from an Amazon S3 file or folder.

updateDataset_name :: Lens' UpdateDataset Text

The name of the dataset to be updated.

updateDataset_input :: Lens' UpdateDataset Input

Undocumented member.

Destructuring the Response

data UpdateDatasetResponse

See: newUpdateDatasetResponse smart constructor.

Constructors

UpdateDatasetResponse'

Fields

httpStatus :: Int

The response's http status code.

name :: Text

The name of the dataset that you updated.

newUpdateDatasetResponse

Arguments

:: Int
$sel:httpStatus:UpdateDatasetResponse'

-> Text
$sel:name:UpdateDataset'

-> UpdateDatasetResponse

Create a value of UpdateDatasetResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:httpStatus:UpdateDatasetResponse', updateDatasetResponse_httpStatus - The response's http status code.

$sel:name:UpdateDataset', updateDatasetResponse_name - The name of the dataset that you updated.

Response Lenses

updateDatasetResponse_httpStatus :: Lens' UpdateDatasetResponse Int

The response's http status code.

updateDatasetResponse_name :: Lens' UpdateDatasetResponse Text

The name of the dataset that you updated.