Amazonka.DataBrew.UpdateRecipe

Description

Modifies the definition of the LATEST_WORKING version of a DataBrew recipe.

Creating a Request

data UpdateRecipe

See: newUpdateRecipe smart constructor.

Constructors

UpdateRecipe'

Fields

steps :: Maybe [RecipeStep]

One or more steps to be performed by the recipe. Each step consists of an action, and the conditions under which the action should succeed.

description :: Maybe Text

A description of the recipe.

name :: Text

The name of the recipe to be updated.

newUpdateRecipe

Arguments

:: Text
$sel:name:UpdateRecipe'

-> UpdateRecipe

Create a value of UpdateRecipe with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:steps:UpdateRecipe', updateRecipe_steps - One or more steps to be performed by the recipe. Each step consists of an action, and the conditions under which the action should succeed.

$sel:description:UpdateRecipe', updateRecipe_description - A description of the recipe.

$sel:name:UpdateRecipe', updateRecipe_name - The name of the recipe to be updated.

Request Lenses

updateRecipe_steps :: Lens' UpdateRecipe (Maybe [RecipeStep])

One or more steps to be performed by the recipe. Each step consists of an action, and the conditions under which the action should succeed.

updateRecipe_description :: Lens' UpdateRecipe (Maybe Text)

A description of the recipe.

updateRecipe_name :: Lens' UpdateRecipe Text

The name of the recipe to be updated.

Destructuring the Response

data UpdateRecipeResponse

See: newUpdateRecipeResponse smart constructor.

Constructors

UpdateRecipeResponse'

Fields

httpStatus :: Int

The response's http status code.

name :: Text

The name of the recipe that was updated.

newUpdateRecipeResponse

Arguments

:: Int
$sel:httpStatus:UpdateRecipeResponse'

-> Text
$sel:name:UpdateRecipe'

-> UpdateRecipeResponse

Create a value of UpdateRecipeResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:httpStatus:UpdateRecipeResponse', updateRecipeResponse_httpStatus - The response's http status code.

$sel:name:UpdateRecipe', updateRecipeResponse_name - The name of the recipe that was updated.

Response Lenses

updateRecipeResponse_httpStatus :: Lens' UpdateRecipeResponse Int

The response's http status code.

updateRecipeResponse_name :: Lens' UpdateRecipeResponse Text

The name of the recipe that was updated.