Amazonka.DataExchange.CreateRevision

Description

This operation creates a revision for a data set.

Creating a Request

data CreateRevision

The request body for CreateRevision.

See: newCreateRevision smart constructor.

Constructors

CreateRevision'

Fields

comment :: Maybe Text

An optional comment about the revision.

tags :: Maybe (HashMap Text Text)

A revision tag is an optional label that you can assign to a revision when you create it. Each tag consists of a key and an optional value, both of which you define. When you use tagging, you can also use tag-based access control in IAM policies to control access to these data sets and revisions.

dataSetId :: Text

The unique identifier for a data set.

newCreateRevision

Arguments

:: Text
$sel:dataSetId:CreateRevision'

-> CreateRevision

Create a value of CreateRevision with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:comment:CreateRevision', createRevision_comment - An optional comment about the revision.

$sel:tags:CreateRevision', createRevision_tags - A revision tag is an optional label that you can assign to a revision when you create it. Each tag consists of a key and an optional value, both of which you define. When you use tagging, you can also use tag-based access control in IAM policies to control access to these data sets and revisions.

$sel:dataSetId:CreateRevision', createRevision_dataSetId - The unique identifier for a data set.

Request Lenses

createRevision_comment :: Lens' CreateRevision (Maybe Text)

An optional comment about the revision.

createRevision_tags :: Lens' CreateRevision (Maybe (HashMap Text Text))

A revision tag is an optional label that you can assign to a revision when you create it. Each tag consists of a key and an optional value, both of which you define. When you use tagging, you can also use tag-based access control in IAM policies to control access to these data sets and revisions.

createRevision_dataSetId :: Lens' CreateRevision Text

The unique identifier for a data set.

Destructuring the Response

data CreateRevisionResponse

See: newCreateRevisionResponse smart constructor.

Constructors

CreateRevisionResponse'

Fields

arn :: Maybe Text

The ARN for the revision.

createdAt :: Maybe POSIX

The date and time that the revision was created, in ISO 8601 format.

sourceId :: Maybe Text

The revision ID of the owned revision corresponding to the entitled revision being viewed. This parameter is returned when a revision owner is viewing the entitled copy of its owned revision.

finalized :: Maybe Bool

To publish a revision to a data set in a product, the revision must first be finalized. Finalizing a revision tells AWS Data Exchange that your changes to the assets in the revision are complete. After it's in this read-only state, you can publish the revision to your products.

Finalized revisions can be published through the AWS Data Exchange console or the AWS Marketplace Catalog API, using the StartChangeSet AWS Marketplace Catalog API action. When using the API, revisions are uniquely identified by their ARN.

dataSetId :: Maybe Text

The unique identifier for the data set associated with this revision.

id :: Maybe Text

The unique identifier for the revision.

updatedAt :: Maybe POSIX

The date and time that the revision was last updated, in ISO 8601 format.

comment :: Maybe Text

An optional comment about the revision.

tags :: Maybe (HashMap Text Text)

The tags for the revision.

httpStatus :: Int

The response's http status code.

newCreateRevisionResponse

Arguments

:: Int
$sel:httpStatus:CreateRevisionResponse'

-> CreateRevisionResponse

Create a value of CreateRevisionResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:arn:CreateRevisionResponse', createRevisionResponse_arn - The ARN for the revision.

$sel:createdAt:CreateRevisionResponse', createRevisionResponse_createdAt - The date and time that the revision was created, in ISO 8601 format.

$sel:sourceId:CreateRevisionResponse', createRevisionResponse_sourceId - The revision ID of the owned revision corresponding to the entitled revision being viewed. This parameter is returned when a revision owner is viewing the entitled copy of its owned revision.

$sel:finalized:CreateRevisionResponse', createRevisionResponse_finalized - To publish a revision to a data set in a product, the revision must first be finalized. Finalizing a revision tells AWS Data Exchange that your changes to the assets in the revision are complete. After it's in this read-only state, you can publish the revision to your products.

Finalized revisions can be published through the AWS Data Exchange console or the AWS Marketplace Catalog API, using the StartChangeSet AWS Marketplace Catalog API action. When using the API, revisions are uniquely identified by their ARN.

$sel:dataSetId:CreateRevision', createRevisionResponse_dataSetId - The unique identifier for the data set associated with this revision.

$sel:id:CreateRevisionResponse', createRevisionResponse_id - The unique identifier for the revision.

$sel:updatedAt:CreateRevisionResponse', createRevisionResponse_updatedAt - The date and time that the revision was last updated, in ISO 8601 format.

$sel:comment:CreateRevision', createRevisionResponse_comment - An optional comment about the revision.

$sel:tags:CreateRevision', createRevisionResponse_tags - The tags for the revision.

$sel:httpStatus:CreateRevisionResponse', createRevisionResponse_httpStatus - The response's http status code.

Response Lenses

createRevisionResponse_arn :: Lens' CreateRevisionResponse (Maybe Text)

The ARN for the revision.

createRevisionResponse_createdAt :: Lens' CreateRevisionResponse (Maybe UTCTime)

The date and time that the revision was created, in ISO 8601 format.

createRevisionResponse_sourceId :: Lens' CreateRevisionResponse (Maybe Text)

The revision ID of the owned revision corresponding to the entitled revision being viewed. This parameter is returned when a revision owner is viewing the entitled copy of its owned revision.

createRevisionResponse_finalized :: Lens' CreateRevisionResponse (Maybe Bool)

To publish a revision to a data set in a product, the revision must first be finalized. Finalizing a revision tells AWS Data Exchange that your changes to the assets in the revision are complete. After it's in this read-only state, you can publish the revision to your products.

Finalized revisions can be published through the AWS Data Exchange console or the AWS Marketplace Catalog API, using the StartChangeSet AWS Marketplace Catalog API action. When using the API, revisions are uniquely identified by their ARN.

createRevisionResponse_dataSetId :: Lens' CreateRevisionResponse (Maybe Text)

The unique identifier for the data set associated with this revision.

createRevisionResponse_id :: Lens' CreateRevisionResponse (Maybe Text)

The unique identifier for the revision.

createRevisionResponse_updatedAt :: Lens' CreateRevisionResponse (Maybe UTCTime)

The date and time that the revision was last updated, in ISO 8601 format.

createRevisionResponse_comment :: Lens' CreateRevisionResponse (Maybe Text)

An optional comment about the revision.

createRevisionResponse_tags :: Lens' CreateRevisionResponse (Maybe (HashMap Text Text))

The tags for the revision.

createRevisionResponse_httpStatus :: Lens' CreateRevisionResponse Int

The response's http status code.