Amazonka.DataExchange.Lens

Description

Operations

GetEventAction

getEventAction_eventActionId :: Lens' GetEventAction Text

The unique identifier for the event action.

getEventActionResponse_event :: Lens' GetEventActionResponse (Maybe Event)

What occurs to start an action.

getEventActionResponse_arn :: Lens' GetEventActionResponse (Maybe Text)

The ARN for the event action.

getEventActionResponse_createdAt :: Lens' GetEventActionResponse (Maybe UTCTime)

The date and time that the event action was created, in ISO 8601 format.

getEventActionResponse_action :: Lens' GetEventActionResponse (Maybe Action)

What occurs after a certain event.

getEventActionResponse_id :: Lens' GetEventActionResponse (Maybe Text)

The unique identifier for the event action.

getEventActionResponse_updatedAt :: Lens' GetEventActionResponse (Maybe UTCTime)

The date and time that the event action was last updated, in ISO 8601 format.

getEventActionResponse_httpStatus :: Lens' GetEventActionResponse Int

The response's http status code.

CreateRevision

createRevision_comment :: Lens' CreateRevision (Maybe Text)

An optional comment about the revision.

createRevision_tags :: Lens' CreateRevision (Maybe (HashMap Text Text))

A revision tag is an optional label that you can assign to a revision when you create it. Each tag consists of a key and an optional value, both of which you define. When you use tagging, you can also use tag-based access control in IAM policies to control access to these data sets and revisions.

createRevision_dataSetId :: Lens' CreateRevision Text

The unique identifier for a data set.

createRevisionResponse_arn :: Lens' CreateRevisionResponse (Maybe Text)

The ARN for the revision.

createRevisionResponse_createdAt :: Lens' CreateRevisionResponse (Maybe UTCTime)

The date and time that the revision was created, in ISO 8601 format.

createRevisionResponse_sourceId :: Lens' CreateRevisionResponse (Maybe Text)

The revision ID of the owned revision corresponding to the entitled revision being viewed. This parameter is returned when a revision owner is viewing the entitled copy of its owned revision.

createRevisionResponse_finalized :: Lens' CreateRevisionResponse (Maybe Bool)

To publish a revision to a data set in a product, the revision must first be finalized. Finalizing a revision tells AWS Data Exchange that your changes to the assets in the revision are complete. After it's in this read-only state, you can publish the revision to your products.

Finalized revisions can be published through the AWS Data Exchange console or the AWS Marketplace Catalog API, using the StartChangeSet AWS Marketplace Catalog API action. When using the API, revisions are uniquely identified by their ARN.

createRevisionResponse_dataSetId :: Lens' CreateRevisionResponse (Maybe Text)

The unique identifier for the data set associated with this revision.

createRevisionResponse_id :: Lens' CreateRevisionResponse (Maybe Text)

The unique identifier for the revision.

createRevisionResponse_updatedAt :: Lens' CreateRevisionResponse (Maybe UTCTime)

The date and time that the revision was last updated, in ISO 8601 format.

createRevisionResponse_comment :: Lens' CreateRevisionResponse (Maybe Text)

An optional comment about the revision.

createRevisionResponse_tags :: Lens' CreateRevisionResponse (Maybe (HashMap Text Text))

The tags for the revision.

createRevisionResponse_httpStatus :: Lens' CreateRevisionResponse Int

The response's http status code.

ListRevisionAssets

listRevisionAssets_nextToken :: Lens' ListRevisionAssets (Maybe Text)

The token value retrieved from a previous call to access the next page of results.

listRevisionAssets_maxResults :: Lens' ListRevisionAssets (Maybe Natural)

The maximum number of results returned by a single call.

listRevisionAssets_revisionId :: Lens' ListRevisionAssets Text

The unique identifier for a revision.

listRevisionAssets_dataSetId :: Lens' ListRevisionAssets Text

The unique identifier for a data set.

listRevisionAssetsResponse_nextToken :: Lens' ListRevisionAssetsResponse (Maybe Text)

The token value retrieved from a previous call to access the next page of results.

listRevisionAssetsResponse_assets :: Lens' ListRevisionAssetsResponse (Maybe [AssetEntry])

The asset objects listed by the request.

listRevisionAssetsResponse_httpStatus :: Lens' ListRevisionAssetsResponse Int

The response's http status code.

DeleteRevision

deleteRevision_revisionId :: Lens' DeleteRevision Text

The unique identifier for a revision.

deleteRevision_dataSetId :: Lens' DeleteRevision Text

The unique identifier for a data set.

UpdateRevision

updateRevision_finalized :: Lens' UpdateRevision (Maybe Bool)

Finalizing a revision tells AWS Data Exchange that your changes to the assets in the revision are complete. After it's in this read-only state, you can publish the revision to your products.

updateRevision_comment :: Lens' UpdateRevision (Maybe Text)

An optional comment about the revision.

updateRevision_revisionId :: Lens' UpdateRevision Text

The unique identifier for a revision.

updateRevision_dataSetId :: Lens' UpdateRevision Text

The unique identifier for a data set.

updateRevisionResponse_arn :: Lens' UpdateRevisionResponse (Maybe Text)

The ARN for the revision.

updateRevisionResponse_createdAt :: Lens' UpdateRevisionResponse (Maybe UTCTime)

The date and time that the revision was created, in ISO 8601 format.

updateRevisionResponse_sourceId :: Lens' UpdateRevisionResponse (Maybe Text)

The revision ID of the owned revision corresponding to the entitled revision being viewed. This parameter is returned when a revision owner is viewing the entitled copy of its owned revision.

updateRevisionResponse_finalized :: Lens' UpdateRevisionResponse (Maybe Bool)

To publish a revision to a data set in a product, the revision must first be finalized. Finalizing a revision tells AWS Data Exchange that changes to the assets in the revision are complete. After it's in this read-only state, you can publish the revision to your products.

Finalized revisions can be published through the AWS Data Exchange console or the AWS Marketplace Catalog API, using the StartChangeSet AWS Marketplace Catalog API action. When using the API, revisions are uniquely identified by their ARN.

updateRevisionResponse_dataSetId :: Lens' UpdateRevisionResponse (Maybe Text)

The unique identifier for the data set associated with this revision.

updateRevisionResponse_id :: Lens' UpdateRevisionResponse (Maybe Text)

The unique identifier for the revision.

updateRevisionResponse_updatedAt :: Lens' UpdateRevisionResponse (Maybe UTCTime)

The date and time that the revision was last updated, in ISO 8601 format.

updateRevisionResponse_comment :: Lens' UpdateRevisionResponse (Maybe Text)

An optional comment about the revision.

updateRevisionResponse_httpStatus :: Lens' UpdateRevisionResponse Int

The response's http status code.

ListTagsForResource

listTagsForResource_resourceArn :: Lens' ListTagsForResource Text

An Amazon Resource Name (ARN) that uniquely identifies an AWS resource.

listTagsForResourceResponse_tags :: Lens' ListTagsForResourceResponse (Maybe (HashMap Text Text))

A label that consists of a customer-defined key and an optional value.

listTagsForResourceResponse_httpStatus :: Lens' ListTagsForResourceResponse Int

The response's http status code.

GetRevision

getRevision_revisionId :: Lens' GetRevision Text

The unique identifier for a revision.

getRevision_dataSetId :: Lens' GetRevision Text

The unique identifier for a data set.

getRevisionResponse_arn :: Lens' GetRevisionResponse (Maybe Text)

The ARN for the revision.

getRevisionResponse_createdAt :: Lens' GetRevisionResponse (Maybe UTCTime)

The date and time that the revision was created, in ISO 8601 format.

getRevisionResponse_sourceId :: Lens' GetRevisionResponse (Maybe Text)

The revision ID of the owned revision corresponding to the entitled revision being viewed. This parameter is returned when a revision owner is viewing the entitled copy of its owned revision.

getRevisionResponse_finalized :: Lens' GetRevisionResponse (Maybe Bool)

To publish a revision to a data set in a product, the revision must first be finalized. Finalizing a revision tells AWS Data Exchange that your changes to the assets in the revision are complete. After it's in this read-only state, you can publish the revision to your products.

Finalized revisions can be published through the AWS Data Exchange console or the AWS Marketplace Catalog API, using the StartChangeSet AWS Marketplace Catalog API action. When using the API, revisions are uniquely identified by their ARN.

getRevisionResponse_dataSetId :: Lens' GetRevisionResponse (Maybe Text)

The unique identifier for the data set associated with this revision.

getRevisionResponse_id :: Lens' GetRevisionResponse (Maybe Text)

The unique identifier for the revision.

getRevisionResponse_updatedAt :: Lens' GetRevisionResponse (Maybe UTCTime)

The date and time that the revision was last updated, in ISO 8601 format.

getRevisionResponse_comment :: Lens' GetRevisionResponse (Maybe Text)

An optional comment about the revision.

getRevisionResponse_tags :: Lens' GetRevisionResponse (Maybe (HashMap Text Text))

The tags for the revision.

getRevisionResponse_httpStatus :: Lens' GetRevisionResponse Int

The response's http status code.

DeleteDataSet

deleteDataSet_dataSetId :: Lens' DeleteDataSet Text

The unique identifier for a data set.

UpdateDataSet

updateDataSet_name :: Lens' UpdateDataSet (Maybe Text)

The name of the data set.

updateDataSet_description :: Lens' UpdateDataSet (Maybe Text)

The description for the data set.

updateDataSet_dataSetId :: Lens' UpdateDataSet Text

The unique identifier for a data set.

updateDataSetResponse_origin :: Lens' UpdateDataSetResponse (Maybe Origin)

A property that defines the data set as OWNED by the account (for providers) or ENTITLED to the account (for subscribers).

updateDataSetResponse_arn :: Lens' UpdateDataSetResponse (Maybe Text)

The ARN for the data set.

updateDataSetResponse_createdAt :: Lens' UpdateDataSetResponse (Maybe UTCTime)

The date and time that the data set was created, in ISO 8601 format.

updateDataSetResponse_sourceId :: Lens' UpdateDataSetResponse (Maybe Text)

The data set ID of the owned data set corresponding to the entitled data set being viewed. This parameter is returned when a data set owner is viewing the entitled copy of its owned data set.

updateDataSetResponse_originDetails :: Lens' UpdateDataSetResponse (Maybe OriginDetails)

If the origin of this data set is ENTITLED, includes the details for the product on AWS Marketplace.

updateDataSetResponse_name :: Lens' UpdateDataSetResponse (Maybe Text)

The name of the data set.

updateDataSetResponse_id :: Lens' UpdateDataSetResponse (Maybe Text)

The unique identifier for the data set.

updateDataSetResponse_assetType :: Lens' UpdateDataSetResponse (Maybe AssetType)

The type of asset that is added to a data set.

updateDataSetResponse_updatedAt :: Lens' UpdateDataSetResponse (Maybe UTCTime)

The date and time that the data set was last updated, in ISO 8601 format.

updateDataSetResponse_description :: Lens' UpdateDataSetResponse (Maybe Text)

The description for the data set.

updateDataSetResponse_httpStatus :: Lens' UpdateDataSetResponse Int

The response's http status code.

CreateJob

createJob_type :: Lens' CreateJob Type

The type of job to be created.

createJob_details :: Lens' CreateJob RequestDetails

The details for the CreateJob request.

createJobResponse_state :: Lens' CreateJobResponse (Maybe State)

The state of the job.

createJobResponse_arn :: Lens' CreateJobResponse (Maybe Text)

The ARN for the job.

createJobResponse_createdAt :: Lens' CreateJobResponse (Maybe UTCTime)

The date and time that the job was created, in ISO 8601 format.

createJobResponse_details :: Lens' CreateJobResponse (Maybe ResponseDetails)

Details about the job.

createJobResponse_id :: Lens' CreateJobResponse (Maybe Text)

The unique identifier for the job.

createJobResponse_type :: Lens' CreateJobResponse (Maybe Type)

The job type.

createJobResponse_updatedAt :: Lens' CreateJobResponse (Maybe UTCTime)

The date and time that the job was last updated, in ISO 8601 format.

createJobResponse_errors :: Lens' CreateJobResponse (Maybe [JobError])

The errors associated with jobs.

createJobResponse_httpStatus :: Lens' CreateJobResponse Int

The response's http status code.

ListEventActions

listEventActions_eventSourceId :: Lens' ListEventActions (Maybe Text)

The unique identifier for the event source.

listEventActions_nextToken :: Lens' ListEventActions (Maybe Text)

The token value retrieved from a previous call to access the next page of results.

listEventActions_maxResults :: Lens' ListEventActions (Maybe Natural)

The maximum number of results returned by a single call.

listEventActionsResponse_eventActions :: Lens' ListEventActionsResponse (Maybe [EventActionEntry])

The event action objects listed by the request.

listEventActionsResponse_nextToken :: Lens' ListEventActionsResponse (Maybe Text)

The token value retrieved from a previous call to access the next page of results.

listEventActionsResponse_httpStatus :: Lens' ListEventActionsResponse Int

The response's http status code.

GetAsset

getAsset_revisionId :: Lens' GetAsset Text

The unique identifier for a revision.

getAsset_assetId :: Lens' GetAsset Text

The unique identifier for an asset.

getAsset_dataSetId :: Lens' GetAsset Text

The unique identifier for a data set.

getAssetResponse_arn :: Lens' GetAssetResponse (Maybe Text)

The ARN for the asset.

getAssetResponse_createdAt :: Lens' GetAssetResponse (Maybe UTCTime)

The date and time that the asset was created, in ISO 8601 format.

getAssetResponse_sourceId :: Lens' GetAssetResponse (Maybe Text)

The asset ID of the owned asset corresponding to the entitled asset being viewed. This parameter is returned when an asset owner is viewing the entitled copy of its owned asset.

getAssetResponse_dataSetId :: Lens' GetAssetResponse (Maybe Text)

The unique identifier for the data set associated with this asset.

getAssetResponse_name :: Lens' GetAssetResponse (Maybe Text)

The name of the asset. When importing from Amazon S3, the S3 object key is used as the asset name. When exporting to Amazon S3, the asset name is used as default target S3 object key.

getAssetResponse_assetDetails :: Lens' GetAssetResponse (Maybe AssetDetails)

Information about the asset.

getAssetResponse_id :: Lens' GetAssetResponse (Maybe Text)

The unique identifier for the asset.

getAssetResponse_assetType :: Lens' GetAssetResponse (Maybe AssetType)

The type of asset that is added to a data set.

getAssetResponse_updatedAt :: Lens' GetAssetResponse (Maybe UTCTime)

The date and time that the asset was last updated, in ISO 8601 format.

getAssetResponse_revisionId :: Lens' GetAssetResponse (Maybe Text)

The unique identifier for the revision associated with this asset.

getAssetResponse_httpStatus :: Lens' GetAssetResponse Int

The response's http status code.

ListJobs

listJobs_nextToken :: Lens' ListJobs (Maybe Text)

The token value retrieved from a previous call to access the next page of results.

listJobs_dataSetId :: Lens' ListJobs (Maybe Text)

The unique identifier for a data set.

listJobs_maxResults :: Lens' ListJobs (Maybe Natural)

The maximum number of results returned by a single call.

listJobs_revisionId :: Lens' ListJobs (Maybe Text)

The unique identifier for a revision.

listJobsResponse_nextToken :: Lens' ListJobsResponse (Maybe Text)

The token value retrieved from a previous call to access the next page of results.

listJobsResponse_jobs :: Lens' ListJobsResponse (Maybe [JobEntry])

The jobs listed by the request.

listJobsResponse_httpStatus :: Lens' ListJobsResponse Int

The response's http status code.

CreateDataSet

createDataSet_tags :: Lens' CreateDataSet (Maybe (HashMap Text Text))

A data set tag is an optional label that you can assign to a data set when you create it. Each tag consists of a key and an optional value, both of which you define. When you use tagging, you can also use tag-based access control in IAM policies to control access to these data sets and revisions.

createDataSet_assetType :: Lens' CreateDataSet AssetType

The type of asset that is added to a data set.

createDataSet_description :: Lens' CreateDataSet Text

A description for the data set. This value can be up to 16,348 characters long.

createDataSet_name :: Lens' CreateDataSet Text

The name of the data set.

createDataSetResponse_origin :: Lens' CreateDataSetResponse (Maybe Origin)

A property that defines the data set as OWNED by the account (for providers) or ENTITLED to the account (for subscribers).

createDataSetResponse_arn :: Lens' CreateDataSetResponse (Maybe Text)

The ARN for the data set.

createDataSetResponse_createdAt :: Lens' CreateDataSetResponse (Maybe UTCTime)

The date and time that the data set was created, in ISO 8601 format.

createDataSetResponse_sourceId :: Lens' CreateDataSetResponse (Maybe Text)

The data set ID of the owned data set corresponding to the entitled data set being viewed. This parameter is returned when a data set owner is viewing the entitled copy of its owned data set.

createDataSetResponse_originDetails :: Lens' CreateDataSetResponse (Maybe OriginDetails)

If the origin of this data set is ENTITLED, includes the details for the product on AWS Marketplace.

createDataSetResponse_name :: Lens' CreateDataSetResponse (Maybe Text)

The name of the data set.

createDataSetResponse_id :: Lens' CreateDataSetResponse (Maybe Text)

The unique identifier for the data set.

createDataSetResponse_assetType :: Lens' CreateDataSetResponse (Maybe AssetType)

The type of asset that is added to a data set.

createDataSetResponse_updatedAt :: Lens' CreateDataSetResponse (Maybe UTCTime)

The date and time that the data set was last updated, in ISO 8601 format.

createDataSetResponse_description :: Lens' CreateDataSetResponse (Maybe Text)

The description for the data set.

createDataSetResponse_tags :: Lens' CreateDataSetResponse (Maybe (HashMap Text Text))

The tags for the data set.

createDataSetResponse_httpStatus :: Lens' CreateDataSetResponse Int

The response's http status code.

DeleteAsset

deleteAsset_revisionId :: Lens' DeleteAsset Text

The unique identifier for a revision.

deleteAsset_assetId :: Lens' DeleteAsset Text

The unique identifier for an asset.

deleteAsset_dataSetId :: Lens' DeleteAsset Text

The unique identifier for a data set.

UpdateAsset

updateAsset_revisionId :: Lens' UpdateAsset Text

The unique identifier for a revision.

updateAsset_assetId :: Lens' UpdateAsset Text

The unique identifier for an asset.

updateAsset_dataSetId :: Lens' UpdateAsset Text

The unique identifier for a data set.

updateAsset_name :: Lens' UpdateAsset Text

The name of the asset. When importing from Amazon S3, the S3 object key is used as the asset name. When exporting to Amazon S3, the asset name is used as default target S3 object key.

updateAssetResponse_arn :: Lens' UpdateAssetResponse (Maybe Text)

The ARN for the asset.

updateAssetResponse_createdAt :: Lens' UpdateAssetResponse (Maybe UTCTime)

The date and time that the asset was created, in ISO 8601 format.

updateAssetResponse_sourceId :: Lens' UpdateAssetResponse (Maybe Text)

The asset ID of the owned asset corresponding to the entitled asset being viewed. This parameter is returned when an asset owner is viewing the entitled copy of its owned asset.

updateAssetResponse_dataSetId :: Lens' UpdateAssetResponse (Maybe Text)

The unique identifier for the data set associated with this asset.

updateAssetResponse_name :: Lens' UpdateAssetResponse (Maybe Text)

The name of the asset. When importing from Amazon S3, the S3 object key is used as the asset name. When exporting to Amazon S3, the asset name is used as default target S3 object key.

updateAssetResponse_assetDetails :: Lens' UpdateAssetResponse (Maybe AssetDetails)

Information about the asset.

updateAssetResponse_id :: Lens' UpdateAssetResponse (Maybe Text)

The unique identifier for the asset.

updateAssetResponse_assetType :: Lens' UpdateAssetResponse (Maybe AssetType)

The type of asset that is added to a data set.

updateAssetResponse_updatedAt :: Lens' UpdateAssetResponse (Maybe UTCTime)

The date and time that the asset was last updated, in ISO 8601 format.

updateAssetResponse_revisionId :: Lens' UpdateAssetResponse (Maybe Text)

The unique identifier for the revision associated with this asset.

updateAssetResponse_httpStatus :: Lens' UpdateAssetResponse Int

The response's http status code.

GetJob

getJob_jobId :: Lens' GetJob Text

The unique identifier for a job.

getJobResponse_state :: Lens' GetJobResponse (Maybe State)

The state of the job.

getJobResponse_arn :: Lens' GetJobResponse (Maybe Text)

The ARN for the job.

getJobResponse_createdAt :: Lens' GetJobResponse (Maybe UTCTime)

The date and time that the job was created, in ISO 8601 format.

getJobResponse_details :: Lens' GetJobResponse (Maybe ResponseDetails)

Details about the job.

getJobResponse_id :: Lens' GetJobResponse (Maybe Text)

The unique identifier for the job.

getJobResponse_type :: Lens' GetJobResponse (Maybe Type)

The job type.

getJobResponse_updatedAt :: Lens' GetJobResponse (Maybe UTCTime)

The date and time that the job was last updated, in ISO 8601 format.

getJobResponse_errors :: Lens' GetJobResponse (Maybe [JobError])

The errors associated with jobs.

getJobResponse_httpStatus :: Lens' GetJobResponse Int

The response's http status code.

GetDataSet

getDataSet_dataSetId :: Lens' GetDataSet Text

The unique identifier for a data set.

getDataSetResponse_origin :: Lens' GetDataSetResponse (Maybe Origin)

A property that defines the data set as OWNED by the account (for providers) or ENTITLED to the account (for subscribers).

getDataSetResponse_arn :: Lens' GetDataSetResponse (Maybe Text)

The ARN for the data set.

getDataSetResponse_createdAt :: Lens' GetDataSetResponse (Maybe UTCTime)

The date and time that the data set was created, in ISO 8601 format.

getDataSetResponse_sourceId :: Lens' GetDataSetResponse (Maybe Text)

The data set ID of the owned data set corresponding to the entitled data set being viewed. This parameter is returned when a data set owner is viewing the entitled copy of its owned data set.

getDataSetResponse_originDetails :: Lens' GetDataSetResponse (Maybe OriginDetails)

If the origin of this data set is ENTITLED, includes the details for the product on AWS Marketplace.

getDataSetResponse_name :: Lens' GetDataSetResponse (Maybe Text)

The name of the data set.

getDataSetResponse_id :: Lens' GetDataSetResponse (Maybe Text)

The unique identifier for the data set.

getDataSetResponse_assetType :: Lens' GetDataSetResponse (Maybe AssetType)

The type of asset that is added to a data set.

getDataSetResponse_updatedAt :: Lens' GetDataSetResponse (Maybe UTCTime)

The date and time that the data set was last updated, in ISO 8601 format.

getDataSetResponse_description :: Lens' GetDataSetResponse (Maybe Text)

The description for the data set.

getDataSetResponse_tags :: Lens' GetDataSetResponse (Maybe (HashMap Text Text))

The tags for the data set.

getDataSetResponse_httpStatus :: Lens' GetDataSetResponse Int

The response's http status code.

StartJob

startJob_jobId :: Lens' StartJob Text

The unique identifier for a job.

startJobResponse_httpStatus :: Lens' StartJobResponse Int

The response's http status code.

TagResource

tagResource_resourceArn :: Lens' TagResource Text

An Amazon Resource Name (ARN) that uniquely identifies an AWS resource.

tagResource_tags :: Lens' TagResource (HashMap Text Text)

A label that consists of a customer-defined key and an optional value.

ListDataSetRevisions

listDataSetRevisions_nextToken :: Lens' ListDataSetRevisions (Maybe Text)

The token value retrieved from a previous call to access the next page of results.

listDataSetRevisions_maxResults :: Lens' ListDataSetRevisions (Maybe Natural)

The maximum number of results returned by a single call.

listDataSetRevisions_dataSetId :: Lens' ListDataSetRevisions Text

The unique identifier for a data set.

listDataSetRevisionsResponse_nextToken :: Lens' ListDataSetRevisionsResponse (Maybe Text)

The token value retrieved from a previous call to access the next page of results.

listDataSetRevisionsResponse_revisions :: Lens' ListDataSetRevisionsResponse (Maybe [RevisionEntry])

The asset objects listed by the request.

listDataSetRevisionsResponse_httpStatus :: Lens' ListDataSetRevisionsResponse Int

The response's http status code.

DeleteEventAction

deleteEventAction_eventActionId :: Lens' DeleteEventAction Text

The unique identifier for the event action.

UpdateEventAction

updateEventAction_action :: Lens' UpdateEventAction (Maybe Action)

What occurs after a certain event.

updateEventAction_eventActionId :: Lens' UpdateEventAction Text

The unique identifier for the event action.

updateEventActionResponse_event :: Lens' UpdateEventActionResponse (Maybe Event)

What occurs to start an action.

updateEventActionResponse_arn :: Lens' UpdateEventActionResponse (Maybe Text)

The ARN for the event action.

updateEventActionResponse_createdAt :: Lens' UpdateEventActionResponse (Maybe UTCTime)

The date and time that the event action was created, in ISO 8601 format.

updateEventActionResponse_action :: Lens' UpdateEventActionResponse (Maybe Action)

What occurs after a certain event.

updateEventActionResponse_id :: Lens' UpdateEventActionResponse (Maybe Text)

The unique identifier for the event action.

updateEventActionResponse_updatedAt :: Lens' UpdateEventActionResponse (Maybe UTCTime)

The date and time that the event action was last updated, in ISO 8601 format.

updateEventActionResponse_httpStatus :: Lens' UpdateEventActionResponse Int

The response's http status code.

UntagResource

untagResource_tagKeys :: Lens' UntagResource [Text]

The key tags.

untagResource_resourceArn :: Lens' UntagResource Text

An Amazon Resource Name (ARN) that uniquely identifies an AWS resource.

ListDataSets

listDataSets_origin :: Lens' ListDataSets (Maybe Text)

A property that defines the data set as OWNED by the account (for providers) or ENTITLED to the account (for subscribers).

listDataSets_nextToken :: Lens' ListDataSets (Maybe Text)

The token value retrieved from a previous call to access the next page of results.

listDataSets_maxResults :: Lens' ListDataSets (Maybe Natural)

The maximum number of results returned by a single call.

listDataSetsResponse_nextToken :: Lens' ListDataSetsResponse (Maybe Text)

The token value retrieved from a previous call to access the next page of results.

listDataSetsResponse_dataSets :: Lens' ListDataSetsResponse (Maybe [DataSetEntry])

The data set objects listed by the request.

listDataSetsResponse_httpStatus :: Lens' ListDataSetsResponse Int

The response's http status code.

CreateEventAction

createEventAction_action :: Lens' CreateEventAction Action

What occurs after a certain event.

createEventAction_event :: Lens' CreateEventAction Event

What occurs to start an action.

createEventActionResponse_event :: Lens' CreateEventActionResponse (Maybe Event)

What occurs to start an action.

createEventActionResponse_arn :: Lens' CreateEventActionResponse (Maybe Text)

The ARN for the event action.

createEventActionResponse_createdAt :: Lens' CreateEventActionResponse (Maybe UTCTime)

The date and time that the event action was created, in ISO 8601 format.

createEventActionResponse_action :: Lens' CreateEventActionResponse (Maybe Action)

What occurs after a certain event.

createEventActionResponse_id :: Lens' CreateEventActionResponse (Maybe Text)

The unique identifier for the event action.

createEventActionResponse_updatedAt :: Lens' CreateEventActionResponse (Maybe UTCTime)

The date and time that the event action was last updated, in ISO 8601 format.

createEventActionResponse_httpStatus :: Lens' CreateEventActionResponse Int

The response's http status code.

CancelJob

cancelJob_jobId :: Lens' CancelJob Text

The unique identifier for a job.

Types

Action

action_exportRevisionToS3 :: Lens' Action (Maybe AutoExportRevisionToS3RequestDetails)

Details for the export revision to Amazon S3 action.

AssetDestinationEntry

assetDestinationEntry_key :: Lens' AssetDestinationEntry (Maybe Text)

The name of the object in Amazon S3 for the asset.

assetDestinationEntry_bucket :: Lens' AssetDestinationEntry Text

The S3 bucket that is the destination for the asset.

assetDestinationEntry_assetId :: Lens' AssetDestinationEntry Text

The unique identifier for the asset.

AssetDetails

assetDetails_s3SnapshotAsset :: Lens' AssetDetails (Maybe S3SnapshotAsset)

The S3 object that is the asset.

assetDetails_redshiftDataShareAsset :: Lens' AssetDetails (Maybe RedshiftDataShareAsset)

The Amazon Redshift datashare that is the asset.

AssetEntry

assetEntry_sourceId :: Lens' AssetEntry (Maybe Text)

The asset ID of the owned asset corresponding to the entitled asset being viewed. This parameter is returned when an asset owner is viewing the entitled copy of its owned asset.

assetEntry_assetType :: Lens' AssetEntry AssetType

The type of asset that is added to a data set.

assetEntry_createdAt :: Lens' AssetEntry UTCTime

The date and time that the asset was created, in ISO 8601 format.

assetEntry_dataSetId :: Lens' AssetEntry Text

The unique identifier for the data set associated with this asset.

assetEntry_id :: Lens' AssetEntry Text

The unique identifier for the asset.

assetEntry_arn :: Lens' AssetEntry Text

The ARN for the asset.

assetEntry_assetDetails :: Lens' AssetEntry AssetDetails

Information about the asset.

assetEntry_updatedAt :: Lens' AssetEntry UTCTime

The date and time that the asset was last updated, in ISO 8601 format.

assetEntry_revisionId :: Lens' AssetEntry Text

The unique identifier for the revision associated with this asset.

assetEntry_name :: Lens' AssetEntry Text

The name of the asset. When importing from Amazon S3, the S3 object key is used as the asset name. When exporting to Amazon S3, the asset name is used as default target S3 object key.

AssetSourceEntry

assetSourceEntry_bucket :: Lens' AssetSourceEntry Text

The S3 bucket that's part of the source of the asset.

assetSourceEntry_key :: Lens' AssetSourceEntry Text

The name of the object in Amazon S3 for the asset.

AutoExportRevisionDestinationEntry

autoExportRevisionDestinationEntry_keyPattern :: Lens' AutoExportRevisionDestinationEntry (Maybe Text)

A string representing the pattern for generated names of the individual assets in the revision. For more information about key patterns, see Key patterns when exporting revisions.

autoExportRevisionDestinationEntry_bucket :: Lens' AutoExportRevisionDestinationEntry Text

The S3 bucket that is the destination for the event action.

AutoExportRevisionToS3RequestDetails

autoExportRevisionToS3RequestDetails_encryption :: Lens' AutoExportRevisionToS3RequestDetails (Maybe ExportServerSideEncryption)

Encryption configuration for the auto export job.

autoExportRevisionToS3RequestDetails_revisionDestination :: Lens' AutoExportRevisionToS3RequestDetails AutoExportRevisionDestinationEntry

A revision destination is the Amazon S3 bucket folder destination to where the export will be sent.

DataSetEntry

dataSetEntry_sourceId :: Lens' DataSetEntry (Maybe Text)

The data set ID of the owned data set corresponding to the entitled data set being viewed. This parameter is returned when a data set owner is viewing the entitled copy of its owned data set.

dataSetEntry_originDetails :: Lens' DataSetEntry (Maybe OriginDetails)

If the origin of this data set is ENTITLED, includes the details for the product on AWS Marketplace.

dataSetEntry_origin :: Lens' DataSetEntry Origin

A property that defines the data set as OWNED by the account (for providers) or ENTITLED to the account (for subscribers).

dataSetEntry_assetType :: Lens' DataSetEntry AssetType

The type of asset that is added to a data set.

dataSetEntry_description :: Lens' DataSetEntry Text

The description for the data set.

dataSetEntry_createdAt :: Lens' DataSetEntry UTCTime

The date and time that the data set was created, in ISO 8601 format.

dataSetEntry_id :: Lens' DataSetEntry Text

The unique identifier for the data set.

dataSetEntry_arn :: Lens' DataSetEntry Text

The ARN for the data set.

dataSetEntry_updatedAt :: Lens' DataSetEntry UTCTime

The date and time that the data set was last updated, in ISO 8601 format.

dataSetEntry_name :: Lens' DataSetEntry Text

The name of the data set.

Details

details_importAssetFromSignedUrlJobErrorDetails :: Lens' Details (Maybe ImportAssetFromSignedUrlJobErrorDetails)

Information about the job error.

details_importAssetsFromS3JobErrorDetails :: Lens' Details (Maybe [AssetSourceEntry])

Information about the job error.

Event

event_revisionPublished :: Lens' Event (Maybe RevisionPublished)

What occurs to start the revision publish action.

EventActionEntry

eventActionEntry_action :: Lens' EventActionEntry Action

What occurs after a certain event.

eventActionEntry_createdAt :: Lens' EventActionEntry UTCTime

The date and time that the event action was created, in ISO 8601 format.

eventActionEntry_event :: Lens' EventActionEntry Event

What occurs to start an action.

eventActionEntry_id :: Lens' EventActionEntry Text

The unique identifier for the event action.

eventActionEntry_arn :: Lens' EventActionEntry Text

The Amazon Resource Name (ARN) for the event action.

eventActionEntry_updatedAt :: Lens' EventActionEntry UTCTime

The date and time that the event action was last updated, in ISO 8601 format.

ExportAssetToSignedUrlRequestDetails

exportAssetToSignedUrlRequestDetails_dataSetId :: Lens' ExportAssetToSignedUrlRequestDetails Text

The unique identifier for the data set associated with this export job.

exportAssetToSignedUrlRequestDetails_assetId :: Lens' ExportAssetToSignedUrlRequestDetails Text

The unique identifier for the asset that is exported to a signed URL.

exportAssetToSignedUrlRequestDetails_revisionId :: Lens' ExportAssetToSignedUrlRequestDetails Text

The unique identifier for the revision associated with this export request.

ExportAssetToSignedUrlResponseDetails

exportAssetToSignedUrlResponseDetails_signedUrl :: Lens' ExportAssetToSignedUrlResponseDetails (Maybe Text)

The signed URL for the export request.

exportAssetToSignedUrlResponseDetails_signedUrlExpiresAt :: Lens' ExportAssetToSignedUrlResponseDetails (Maybe UTCTime)

The date and time that the signed URL expires, in ISO 8601 format.

exportAssetToSignedUrlResponseDetails_dataSetId :: Lens' ExportAssetToSignedUrlResponseDetails Text

The unique identifier for the data set associated with this export job.

exportAssetToSignedUrlResponseDetails_assetId :: Lens' ExportAssetToSignedUrlResponseDetails Text

The unique identifier for the asset associated with this export job.

exportAssetToSignedUrlResponseDetails_revisionId :: Lens' ExportAssetToSignedUrlResponseDetails Text

The unique identifier for the revision associated with this export response.

ExportAssetsToS3RequestDetails

exportAssetsToS3RequestDetails_encryption :: Lens' ExportAssetsToS3RequestDetails (Maybe ExportServerSideEncryption)

Encryption configuration for the export job.

exportAssetsToS3RequestDetails_assetDestinations :: Lens' ExportAssetsToS3RequestDetails [AssetDestinationEntry]

The destination for the asset.

exportAssetsToS3RequestDetails_dataSetId :: Lens' ExportAssetsToS3RequestDetails Text

The unique identifier for the data set associated with this export job.

exportAssetsToS3RequestDetails_revisionId :: Lens' ExportAssetsToS3RequestDetails Text

The unique identifier for the revision associated with this export request.

ExportAssetsToS3ResponseDetails

exportAssetsToS3ResponseDetails_encryption :: Lens' ExportAssetsToS3ResponseDetails (Maybe ExportServerSideEncryption)

Encryption configuration of the export job.

exportAssetsToS3ResponseDetails_assetDestinations :: Lens' ExportAssetsToS3ResponseDetails [AssetDestinationEntry]

The destination in Amazon S3 where the asset is exported.

exportAssetsToS3ResponseDetails_dataSetId :: Lens' ExportAssetsToS3ResponseDetails Text

The unique identifier for the data set associated with this export job.

exportAssetsToS3ResponseDetails_revisionId :: Lens' ExportAssetsToS3ResponseDetails Text

The unique identifier for the revision associated with this export response.

ExportRevisionsToS3RequestDetails

exportRevisionsToS3RequestDetails_encryption :: Lens' ExportRevisionsToS3RequestDetails (Maybe ExportServerSideEncryption)

Encryption configuration for the export job.

exportRevisionsToS3RequestDetails_revisionDestinations :: Lens' ExportRevisionsToS3RequestDetails [RevisionDestinationEntry]

The destination for the revision.

exportRevisionsToS3RequestDetails_dataSetId :: Lens' ExportRevisionsToS3RequestDetails Text

The unique identifier for the data set associated with this export job.

ExportRevisionsToS3ResponseDetails

exportRevisionsToS3ResponseDetails_encryption :: Lens' ExportRevisionsToS3ResponseDetails (Maybe ExportServerSideEncryption)

Encryption configuration of the export job.

exportRevisionsToS3ResponseDetails_eventActionArn :: Lens' ExportRevisionsToS3ResponseDetails (Maybe Text)

The Amazon Resource Name (ARN) of the event action.

exportRevisionsToS3ResponseDetails_revisionDestinations :: Lens' ExportRevisionsToS3ResponseDetails [RevisionDestinationEntry]

The destination in Amazon S3 where the revision is exported.

exportRevisionsToS3ResponseDetails_dataSetId :: Lens' ExportRevisionsToS3ResponseDetails Text

The unique identifier for the data set associated with this export job.

ExportServerSideEncryption

exportServerSideEncryption_kmsKeyArn :: Lens' ExportServerSideEncryption (Maybe Text)

The Amazon Resource Name (ARN) of the AWS KMS key you want to use to encrypt the Amazon S3 objects. This parameter is required if you choose aws:kms as an encryption type.

exportServerSideEncryption_type :: Lens' ExportServerSideEncryption ServerSideEncryptionTypes

The type of server side encryption used for encrypting the objects in Amazon S3.

ImportAssetFromSignedUrlJobErrorDetails

importAssetFromSignedUrlJobErrorDetails_assetName :: Lens' ImportAssetFromSignedUrlJobErrorDetails Text

Information about the job error.

ImportAssetFromSignedUrlRequestDetails

importAssetFromSignedUrlRequestDetails_dataSetId :: Lens' ImportAssetFromSignedUrlRequestDetails Text

The unique identifier for the data set associated with this import job.

importAssetFromSignedUrlRequestDetails_md5Hash :: Lens' ImportAssetFromSignedUrlRequestDetails Text

The Base64-encoded Md5 hash for the asset, used to ensure the integrity of the file at that location.

importAssetFromSignedUrlRequestDetails_assetName :: Lens' ImportAssetFromSignedUrlRequestDetails Text

The name of the asset. When importing from Amazon S3, the S3 object key is used as the asset name.

importAssetFromSignedUrlRequestDetails_revisionId :: Lens' ImportAssetFromSignedUrlRequestDetails Text

The unique identifier for the revision associated with this import request.

ImportAssetFromSignedUrlResponseDetails

importAssetFromSignedUrlResponseDetails_signedUrl :: Lens' ImportAssetFromSignedUrlResponseDetails (Maybe Text)

The signed URL.

importAssetFromSignedUrlResponseDetails_signedUrlExpiresAt :: Lens' ImportAssetFromSignedUrlResponseDetails (Maybe UTCTime)

The time and date at which the signed URL expires, in ISO 8601 format.

importAssetFromSignedUrlResponseDetails_md5Hash :: Lens' ImportAssetFromSignedUrlResponseDetails (Maybe Text)

The Base64-encoded Md5 hash for the asset, used to ensure the integrity of the file at that location.

importAssetFromSignedUrlResponseDetails_dataSetId :: Lens' ImportAssetFromSignedUrlResponseDetails Text

The unique identifier for the data set associated with this import job.

importAssetFromSignedUrlResponseDetails_assetName :: Lens' ImportAssetFromSignedUrlResponseDetails Text

The name for the asset associated with this import job.

importAssetFromSignedUrlResponseDetails_revisionId :: Lens' ImportAssetFromSignedUrlResponseDetails Text

The unique identifier for the revision associated with this import response.

ImportAssetsFromRedshiftDataSharesRequestDetails

importAssetsFromRedshiftDataSharesRequestDetails_dataSetId :: Lens' ImportAssetsFromRedshiftDataSharesRequestDetails Text

The unique identifier for the data set associated with this import job.

importAssetsFromRedshiftDataSharesRequestDetails_assetSources :: Lens' ImportAssetsFromRedshiftDataSharesRequestDetails [RedshiftDataShareAssetSourceEntry]

A list of Amazon Redshift datashare assets.

importAssetsFromRedshiftDataSharesRequestDetails_revisionId :: Lens' ImportAssetsFromRedshiftDataSharesRequestDetails Text

The unique identifier for the revision associated with this import job.

ImportAssetsFromRedshiftDataSharesResponseDetails

importAssetsFromRedshiftDataSharesResponseDetails_dataSetId :: Lens' ImportAssetsFromRedshiftDataSharesResponseDetails Text

The unique identifier for the data set associated with this import job.

importAssetsFromRedshiftDataSharesResponseDetails_assetSources :: Lens' ImportAssetsFromRedshiftDataSharesResponseDetails [RedshiftDataShareAssetSourceEntry]

A list of Amazon Redshift datashare asset sources.

importAssetsFromRedshiftDataSharesResponseDetails_revisionId :: Lens' ImportAssetsFromRedshiftDataSharesResponseDetails Text

The unique identifier for the revision associated with this import job.

ImportAssetsFromS3RequestDetails

importAssetsFromS3RequestDetails_dataSetId :: Lens' ImportAssetsFromS3RequestDetails Text

The unique identifier for the data set associated with this import job.

importAssetsFromS3RequestDetails_assetSources :: Lens' ImportAssetsFromS3RequestDetails [AssetSourceEntry]

Is a list of S3 bucket and object key pairs.

importAssetsFromS3RequestDetails_revisionId :: Lens' ImportAssetsFromS3RequestDetails Text

The unique identifier for the revision associated with this import request.

ImportAssetsFromS3ResponseDetails

importAssetsFromS3ResponseDetails_dataSetId :: Lens' ImportAssetsFromS3ResponseDetails Text

The unique identifier for the data set associated with this import job.

importAssetsFromS3ResponseDetails_assetSources :: Lens' ImportAssetsFromS3ResponseDetails [AssetSourceEntry]

Is a list of Amazon S3 bucket and object key pairs.

importAssetsFromS3ResponseDetails_revisionId :: Lens' ImportAssetsFromS3ResponseDetails Text

The unique identifier for the revision associated with this import response.

JobEntry

jobEntry_errors :: Lens' JobEntry (Maybe [JobError])

Errors for jobs.

jobEntry_type :: Lens' JobEntry Type

The job type.

jobEntry_details :: Lens' JobEntry ResponseDetails

Details of the operation to be performed by the job, such as export destination details or import source details.

jobEntry_state :: Lens' JobEntry State

The state of the job.

jobEntry_createdAt :: Lens' JobEntry UTCTime

The date and time that the job was created, in ISO 8601 format.

jobEntry_id :: Lens' JobEntry Text

The unique identifier for the job.

jobEntry_arn :: Lens' JobEntry Text

The ARN for the job.

jobEntry_updatedAt :: Lens' JobEntry UTCTime

The date and time that the job was last updated, in ISO 8601 format.

JobError

jobError_resourceId :: Lens' JobError (Maybe Text)

The unique identifier for the resource related to the error.

jobError_limitName :: Lens' JobError (Maybe JobErrorLimitName)

The name of the limit that was reached.

jobError_resourceType :: Lens' JobError (Maybe JobErrorResourceTypes)

The type of resource related to the error.

jobError_details :: Lens' JobError (Maybe Details)

The details about the job error.

jobError_limitValue :: Lens' JobError (Maybe Double)

The value of the exceeded limit.

jobError_message :: Lens' JobError Text

The message related to the job error.

jobError_code :: Lens' JobError Code

The code for the job error.

OriginDetails

originDetails_productId :: Lens' OriginDetails Text

The product ID of the origin of the data set.

RedshiftDataShareAsset

redshiftDataShareAsset_arn :: Lens' RedshiftDataShareAsset Text

The Amazon Resource Name (ARN) of the datashare asset.

RedshiftDataShareAssetSourceEntry

redshiftDataShareAssetSourceEntry_dataShareArn :: Lens' RedshiftDataShareAssetSourceEntry Text

The Amazon Resource Name (ARN) of the datashare asset.

RequestDetails

requestDetails_exportAssetsToS3 :: Lens' RequestDetails (Maybe ExportAssetsToS3RequestDetails)

Details about the export to Amazon S3 request.

requestDetails_exportRevisionsToS3 :: Lens' RequestDetails (Maybe ExportRevisionsToS3RequestDetails)

Details about the export to Amazon S3 request.

requestDetails_importAssetFromSignedUrl :: Lens' RequestDetails (Maybe ImportAssetFromSignedUrlRequestDetails)

Details about the import from signed URL request.

requestDetails_importAssetsFromRedshiftDataShares :: Lens' RequestDetails (Maybe ImportAssetsFromRedshiftDataSharesRequestDetails)

Details from an import from Amazon Redshift datashare request.

requestDetails_importAssetsFromS3 :: Lens' RequestDetails (Maybe ImportAssetsFromS3RequestDetails)

Details about the import from Amazon S3 request.

requestDetails_exportAssetToSignedUrl :: Lens' RequestDetails (Maybe ExportAssetToSignedUrlRequestDetails)

Details about the export to signed URL request.

ResponseDetails

responseDetails_exportAssetsToS3 :: Lens' ResponseDetails (Maybe ExportAssetsToS3ResponseDetails)

Details for the export to Amazon S3 response.

responseDetails_exportRevisionsToS3 :: Lens' ResponseDetails (Maybe ExportRevisionsToS3ResponseDetails)

Details for the export revisions to Amazon S3 response.

responseDetails_importAssetFromSignedUrl :: Lens' ResponseDetails (Maybe ImportAssetFromSignedUrlResponseDetails)

Details for the import from signed URL response.

responseDetails_importAssetsFromRedshiftDataShares :: Lens' ResponseDetails (Maybe ImportAssetsFromRedshiftDataSharesResponseDetails)

Details from an import from Amazon Redshift datashare response.

responseDetails_importAssetsFromS3 :: Lens' ResponseDetails (Maybe ImportAssetsFromS3ResponseDetails)

Details for the import from Amazon S3 response.

responseDetails_exportAssetToSignedUrl :: Lens' ResponseDetails (Maybe ExportAssetToSignedUrlResponseDetails)

Details for the export to signed URL response.

RevisionDestinationEntry

revisionDestinationEntry_keyPattern :: Lens' RevisionDestinationEntry (Maybe Text)

A string representing the pattern for generated names of the individual assets in the revision. For more information about key patterns, see Key patterns when exporting revisions.

revisionDestinationEntry_bucket :: Lens' RevisionDestinationEntry Text

The S3 bucket that is the destination for the assets in the revision.

revisionDestinationEntry_revisionId :: Lens' RevisionDestinationEntry Text

The unique identifier for the revision.

RevisionEntry

revisionEntry_sourceId :: Lens' RevisionEntry (Maybe Text)

The revision ID of the owned revision corresponding to the entitled revision being viewed. This parameter is returned when a revision owner is viewing the entitled copy of its owned revision.

revisionEntry_finalized :: Lens' RevisionEntry (Maybe Bool)

To publish a revision to a data set in a product, the revision must first be finalized. Finalizing a revision tells AWS Data Exchange that your changes to the assets in the revision are complete. After it's in this read-only state, you can publish the revision to your products.

Finalized revisions can be published through the AWS Data Exchange console or the AWS Marketplace Catalog API, using the StartChangeSet AWS Marketplace Catalog API action. When using the API, revisions are uniquely identified by their ARN.

revisionEntry_comment :: Lens' RevisionEntry (Maybe Text)

An optional comment about the revision.

revisionEntry_createdAt :: Lens' RevisionEntry UTCTime

The date and time that the revision was created, in ISO 8601 format.

revisionEntry_dataSetId :: Lens' RevisionEntry Text

The unique identifier for the data set associated with this revision.

revisionEntry_id :: Lens' RevisionEntry Text

The unique identifier for the revision.

revisionEntry_arn :: Lens' RevisionEntry Text

The ARN for the revision.

revisionEntry_updatedAt :: Lens' RevisionEntry UTCTime

The date and time that the revision was last updated, in ISO 8601 format.

RevisionPublished

revisionPublished_dataSetId :: Lens' RevisionPublished Text

The data set ID of the published revision.

S3SnapshotAsset

s3SnapshotAsset_size :: Lens' S3SnapshotAsset Double

The size of the S3 object that is the object.