Amazonka.DataExchange.Types.AssetEntry

Description

Documentation

data AssetEntry

An asset in AWS Data Exchange is a piece of data. The asset can be a structured data file, an image file, or some other data file that can be stored as an S3 object, or an Amazon Redshift datashare (Preview). When you create an import job for your files, you create an asset in AWS Data Exchange for each of those files.

See: newAssetEntry smart constructor.

Constructors

AssetEntry'

Fields

sourceId :: Maybe Text

The asset ID of the owned asset corresponding to the entitled asset being viewed. This parameter is returned when an asset owner is viewing the entitled copy of its owned asset.

assetType :: AssetType

The type of asset that is added to a data set.

createdAt :: POSIX

The date and time that the asset was created, in ISO 8601 format.

dataSetId :: Text

The unique identifier for the data set associated with this asset.

id :: Text

The unique identifier for the asset.

arn :: Text

The ARN for the asset.

assetDetails :: AssetDetails

Information about the asset.

updatedAt :: POSIX

The date and time that the asset was last updated, in ISO 8601 format.

revisionId :: Text

The unique identifier for the revision associated with this asset.

name :: Text

The name of the asset. When importing from Amazon S3, the S3 object key is used as the asset name. When exporting to Amazon S3, the asset name is used as default target S3 object key.

newAssetEntry

Arguments

:: AssetType
$sel:assetType:AssetEntry'

-> UTCTime
$sel:createdAt:AssetEntry'

-> Text
$sel:dataSetId:AssetEntry'

-> Text
$sel:id:AssetEntry'

-> Text
$sel:arn:AssetEntry'

-> AssetDetails
$sel:assetDetails:AssetEntry'

-> UTCTime
$sel:updatedAt:AssetEntry'

-> Text
$sel:revisionId:AssetEntry'

-> Text
$sel:name:AssetEntry'

-> AssetEntry

Create a value of AssetEntry with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:sourceId:AssetEntry', assetEntry_sourceId - The asset ID of the owned asset corresponding to the entitled asset being viewed. This parameter is returned when an asset owner is viewing the entitled copy of its owned asset.

$sel:assetType:AssetEntry', assetEntry_assetType - The type of asset that is added to a data set.

$sel:createdAt:AssetEntry', assetEntry_createdAt - The date and time that the asset was created, in ISO 8601 format.

$sel:dataSetId:AssetEntry', assetEntry_dataSetId - The unique identifier for the data set associated with this asset.

$sel:id:AssetEntry', assetEntry_id - The unique identifier for the asset.

$sel:arn:AssetEntry', assetEntry_arn - The ARN for the asset.

$sel:assetDetails:AssetEntry', assetEntry_assetDetails - Information about the asset.

$sel:updatedAt:AssetEntry', assetEntry_updatedAt - The date and time that the asset was last updated, in ISO 8601 format.

$sel:revisionId:AssetEntry', assetEntry_revisionId - The unique identifier for the revision associated with this asset.

$sel:name:AssetEntry', assetEntry_name - The name of the asset. When importing from Amazon S3, the S3 object key is used as the asset name. When exporting to Amazon S3, the asset name is used as default target S3 object key.

assetEntry_sourceId :: Lens' AssetEntry (Maybe Text)

The asset ID of the owned asset corresponding to the entitled asset being viewed. This parameter is returned when an asset owner is viewing the entitled copy of its owned asset.

assetEntry_assetType :: Lens' AssetEntry AssetType

The type of asset that is added to a data set.

assetEntry_createdAt :: Lens' AssetEntry UTCTime

The date and time that the asset was created, in ISO 8601 format.

assetEntry_dataSetId :: Lens' AssetEntry Text

The unique identifier for the data set associated with this asset.

assetEntry_id :: Lens' AssetEntry Text

The unique identifier for the asset.

assetEntry_arn :: Lens' AssetEntry Text

The ARN for the asset.

assetEntry_assetDetails :: Lens' AssetEntry AssetDetails

Information about the asset.

assetEntry_updatedAt :: Lens' AssetEntry UTCTime

The date and time that the asset was last updated, in ISO 8601 format.

assetEntry_revisionId :: Lens' AssetEntry Text

The unique identifier for the revision associated with this asset.

assetEntry_name :: Lens' AssetEntry Text

The name of the asset. When importing from Amazon S3, the S3 object key is used as the asset name. When exporting to Amazon S3, the asset name is used as default target S3 object key.