Amazonka.DataExchange.Types.AutoExportRevisionDestinationEntry

Description

Documentation

data AutoExportRevisionDestinationEntry

A revision destination is the Amazon S3 bucket folder destination to where the export will be sent.

See: newAutoExportRevisionDestinationEntry smart constructor.

Constructors

AutoExportRevisionDestinationEntry'

Fields

keyPattern :: Maybe Text

A string representing the pattern for generated names of the individual assets in the revision. For more information about key patterns, see Key patterns when exporting revisions.

bucket :: Text

The S3 bucket that is the destination for the event action.

newAutoExportRevisionDestinationEntry

Arguments

:: Text
$sel:bucket:AutoExportRevisionDestinationEntry'

-> AutoExportRevisionDestinationEntry

Create a value of AutoExportRevisionDestinationEntry with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:keyPattern:AutoExportRevisionDestinationEntry', autoExportRevisionDestinationEntry_keyPattern - A string representing the pattern for generated names of the individual assets in the revision. For more information about key patterns, see Key patterns when exporting revisions.

$sel:bucket:AutoExportRevisionDestinationEntry', autoExportRevisionDestinationEntry_bucket - The S3 bucket that is the destination for the event action.

autoExportRevisionDestinationEntry_keyPattern :: Lens' AutoExportRevisionDestinationEntry (Maybe Text)

A string representing the pattern for generated names of the individual assets in the revision. For more information about key patterns, see Key patterns when exporting revisions.

autoExportRevisionDestinationEntry_bucket :: Lens' AutoExportRevisionDestinationEntry Text

The S3 bucket that is the destination for the event action.