Amazonka.DataExchange.Types.RequestDetails

Description

Documentation

data RequestDetails

The details for the request.

See: newRequestDetails smart constructor.

Constructors

RequestDetails'

Fields

exportAssetsToS3 :: Maybe ExportAssetsToS3RequestDetails

Details about the export to Amazon S3 request.

exportRevisionsToS3 :: Maybe ExportRevisionsToS3RequestDetails

Details about the export to Amazon S3 request.

importAssetFromSignedUrl :: Maybe ImportAssetFromSignedUrlRequestDetails

Details about the import from signed URL request.

importAssetsFromRedshiftDataShares :: Maybe ImportAssetsFromRedshiftDataSharesRequestDetails

Details from an import from Amazon Redshift datashare request.

importAssetsFromS3 :: Maybe ImportAssetsFromS3RequestDetails

Details about the import from Amazon S3 request.

exportAssetToSignedUrl :: Maybe ExportAssetToSignedUrlRequestDetails

Details about the export to signed URL request.

newRequestDetails :: RequestDetails

Create a value of RequestDetails with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:exportAssetsToS3:RequestDetails', requestDetails_exportAssetsToS3 - Details about the export to Amazon S3 request.

$sel:exportRevisionsToS3:RequestDetails', requestDetails_exportRevisionsToS3 - Details about the export to Amazon S3 request.

$sel:importAssetFromSignedUrl:RequestDetails', requestDetails_importAssetFromSignedUrl - Details about the import from signed URL request.

$sel:importAssetsFromRedshiftDataShares:RequestDetails', requestDetails_importAssetsFromRedshiftDataShares - Details from an import from Amazon Redshift datashare request.

$sel:importAssetsFromS3:RequestDetails', requestDetails_importAssetsFromS3 - Details about the import from Amazon S3 request.

$sel:exportAssetToSignedUrl:RequestDetails', requestDetails_exportAssetToSignedUrl - Details about the export to signed URL request.

requestDetails_exportAssetsToS3 :: Lens' RequestDetails (Maybe ExportAssetsToS3RequestDetails)

Details about the export to Amazon S3 request.

requestDetails_exportRevisionsToS3 :: Lens' RequestDetails (Maybe ExportRevisionsToS3RequestDetails)

Details about the export to Amazon S3 request.

requestDetails_importAssetFromSignedUrl :: Lens' RequestDetails (Maybe ImportAssetFromSignedUrlRequestDetails)

Details about the import from signed URL request.

requestDetails_importAssetsFromRedshiftDataShares :: Lens' RequestDetails (Maybe ImportAssetsFromRedshiftDataSharesRequestDetails)

Details from an import from Amazon Redshift datashare request.

requestDetails_importAssetsFromS3 :: Lens' RequestDetails (Maybe ImportAssetsFromS3RequestDetails)

Details about the import from Amazon S3 request.

requestDetails_exportAssetToSignedUrl :: Lens' RequestDetails (Maybe ExportAssetToSignedUrlRequestDetails)

Details about the export to signed URL request.