Amazonka.DataExchange.UpdateAsset

Description

This operation updates an asset.

Creating a Request

data UpdateAsset

The request body for UpdateAsset.

See: newUpdateAsset smart constructor.

Constructors

UpdateAsset'

Fields

revisionId :: Text

The unique identifier for a revision.

assetId :: Text

The unique identifier for an asset.

dataSetId :: Text

The unique identifier for a data set.

name :: Text

The name of the asset. When importing from Amazon S3, the S3 object key is used as the asset name. When exporting to Amazon S3, the asset name is used as default target S3 object key.

newUpdateAsset

Arguments

:: Text
$sel:revisionId:UpdateAsset'

-> Text
$sel:assetId:UpdateAsset'

-> Text
$sel:dataSetId:UpdateAsset'

-> Text
$sel:name:UpdateAsset'

-> UpdateAsset

Create a value of UpdateAsset with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:revisionId:UpdateAsset', updateAsset_revisionId - The unique identifier for a revision.

$sel:assetId:UpdateAsset', updateAsset_assetId - The unique identifier for an asset.

$sel:dataSetId:UpdateAsset', updateAsset_dataSetId - The unique identifier for a data set.

$sel:name:UpdateAsset', updateAsset_name - The name of the asset. When importing from Amazon S3, the S3 object key is used as the asset name. When exporting to Amazon S3, the asset name is used as default target S3 object key.

Request Lenses

updateAsset_revisionId :: Lens' UpdateAsset Text

The unique identifier for a revision.

updateAsset_assetId :: Lens' UpdateAsset Text

The unique identifier for an asset.

updateAsset_dataSetId :: Lens' UpdateAsset Text

The unique identifier for a data set.

updateAsset_name :: Lens' UpdateAsset Text

The name of the asset. When importing from Amazon S3, the S3 object key is used as the asset name. When exporting to Amazon S3, the asset name is used as default target S3 object key.

Destructuring the Response

data UpdateAssetResponse

See: newUpdateAssetResponse smart constructor.

Constructors

UpdateAssetResponse'

Fields

arn :: Maybe Text

The ARN for the asset.

createdAt :: Maybe POSIX

The date and time that the asset was created, in ISO 8601 format.

sourceId :: Maybe Text

The asset ID of the owned asset corresponding to the entitled asset being viewed. This parameter is returned when an asset owner is viewing the entitled copy of its owned asset.

dataSetId :: Maybe Text

The unique identifier for the data set associated with this asset.

name :: Maybe Text

The name of the asset. When importing from Amazon S3, the S3 object key is used as the asset name. When exporting to Amazon S3, the asset name is used as default target S3 object key.

assetDetails :: Maybe AssetDetails

Information about the asset.

id :: Maybe Text

The unique identifier for the asset.

assetType :: Maybe AssetType

The type of asset that is added to a data set.

updatedAt :: Maybe POSIX

The date and time that the asset was last updated, in ISO 8601 format.

revisionId :: Maybe Text

The unique identifier for the revision associated with this asset.

httpStatus :: Int

The response's http status code.

newUpdateAssetResponse

Arguments

:: Int
$sel:httpStatus:UpdateAssetResponse'

-> UpdateAssetResponse

Create a value of UpdateAssetResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:arn:UpdateAssetResponse', updateAssetResponse_arn - The ARN for the asset.

$sel:createdAt:UpdateAssetResponse', updateAssetResponse_createdAt - The date and time that the asset was created, in ISO 8601 format.

$sel:sourceId:UpdateAssetResponse', updateAssetResponse_sourceId - The asset ID of the owned asset corresponding to the entitled asset being viewed. This parameter is returned when an asset owner is viewing the entitled copy of its owned asset.

$sel:dataSetId:UpdateAsset', updateAssetResponse_dataSetId - The unique identifier for the data set associated with this asset.

$sel:name:UpdateAsset', updateAssetResponse_name - The name of the asset. When importing from Amazon S3, the S3 object key is used as the asset name. When exporting to Amazon S3, the asset name is used as default target S3 object key.

$sel:assetDetails:UpdateAssetResponse', updateAssetResponse_assetDetails - Information about the asset.

$sel:id:UpdateAssetResponse', updateAssetResponse_id - The unique identifier for the asset.

$sel:assetType:UpdateAssetResponse', updateAssetResponse_assetType - The type of asset that is added to a data set.

$sel:updatedAt:UpdateAssetResponse', updateAssetResponse_updatedAt - The date and time that the asset was last updated, in ISO 8601 format.

$sel:revisionId:UpdateAsset', updateAssetResponse_revisionId - The unique identifier for the revision associated with this asset.

$sel:httpStatus:UpdateAssetResponse', updateAssetResponse_httpStatus - The response's http status code.

Response Lenses

updateAssetResponse_arn :: Lens' UpdateAssetResponse (Maybe Text)

The ARN for the asset.

updateAssetResponse_createdAt :: Lens' UpdateAssetResponse (Maybe UTCTime)

The date and time that the asset was created, in ISO 8601 format.

updateAssetResponse_sourceId :: Lens' UpdateAssetResponse (Maybe Text)

The asset ID of the owned asset corresponding to the entitled asset being viewed. This parameter is returned when an asset owner is viewing the entitled copy of its owned asset.

updateAssetResponse_dataSetId :: Lens' UpdateAssetResponse (Maybe Text)

The unique identifier for the data set associated with this asset.

updateAssetResponse_name :: Lens' UpdateAssetResponse (Maybe Text)

The name of the asset. When importing from Amazon S3, the S3 object key is used as the asset name. When exporting to Amazon S3, the asset name is used as default target S3 object key.

updateAssetResponse_assetDetails :: Lens' UpdateAssetResponse (Maybe AssetDetails)

Information about the asset.

updateAssetResponse_id :: Lens' UpdateAssetResponse (Maybe Text)

The unique identifier for the asset.

updateAssetResponse_assetType :: Lens' UpdateAssetResponse (Maybe AssetType)

The type of asset that is added to a data set.

updateAssetResponse_updatedAt :: Lens' UpdateAssetResponse (Maybe UTCTime)

The date and time that the asset was last updated, in ISO 8601 format.

updateAssetResponse_revisionId :: Lens' UpdateAssetResponse (Maybe Text)

The unique identifier for the revision associated with this asset.

updateAssetResponse_httpStatus :: Lens' UpdateAssetResponse Int

The response's http status code.