Amazonka.DataExchange.UpdateRevision

Description

This operation updates a revision.

Creating a Request

data UpdateRevision

The request body for UpdateRevision.

See: newUpdateRevision smart constructor.

Constructors

UpdateRevision'

Fields

finalized :: Maybe Bool

Finalizing a revision tells AWS Data Exchange that your changes to the assets in the revision are complete. After it's in this read-only state, you can publish the revision to your products.

comment :: Maybe Text

An optional comment about the revision.

revisionId :: Text

The unique identifier for a revision.

dataSetId :: Text

The unique identifier for a data set.

newUpdateRevision

Arguments

:: Text
$sel:revisionId:UpdateRevision'

-> Text
$sel:dataSetId:UpdateRevision'

-> UpdateRevision

Create a value of UpdateRevision with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:finalized:UpdateRevision', updateRevision_finalized - Finalizing a revision tells AWS Data Exchange that your changes to the assets in the revision are complete. After it's in this read-only state, you can publish the revision to your products.

$sel:comment:UpdateRevision', updateRevision_comment - An optional comment about the revision.

$sel:revisionId:UpdateRevision', updateRevision_revisionId - The unique identifier for a revision.

$sel:dataSetId:UpdateRevision', updateRevision_dataSetId - The unique identifier for a data set.

Request Lenses

updateRevision_finalized :: Lens' UpdateRevision (Maybe Bool)

Finalizing a revision tells AWS Data Exchange that your changes to the assets in the revision are complete. After it's in this read-only state, you can publish the revision to your products.

updateRevision_comment :: Lens' UpdateRevision (Maybe Text)

An optional comment about the revision.

updateRevision_revisionId :: Lens' UpdateRevision Text

The unique identifier for a revision.

updateRevision_dataSetId :: Lens' UpdateRevision Text

The unique identifier for a data set.

Destructuring the Response

data UpdateRevisionResponse

See: newUpdateRevisionResponse smart constructor.

Constructors

UpdateRevisionResponse'

Fields

arn :: Maybe Text

The ARN for the revision.

createdAt :: Maybe POSIX

The date and time that the revision was created, in ISO 8601 format.

sourceId :: Maybe Text

The revision ID of the owned revision corresponding to the entitled revision being viewed. This parameter is returned when a revision owner is viewing the entitled copy of its owned revision.

finalized :: Maybe Bool

To publish a revision to a data set in a product, the revision must first be finalized. Finalizing a revision tells AWS Data Exchange that changes to the assets in the revision are complete. After it's in this read-only state, you can publish the revision to your products.

Finalized revisions can be published through the AWS Data Exchange console or the AWS Marketplace Catalog API, using the StartChangeSet AWS Marketplace Catalog API action. When using the API, revisions are uniquely identified by their ARN.

dataSetId :: Maybe Text

The unique identifier for the data set associated with this revision.

id :: Maybe Text

The unique identifier for the revision.

updatedAt :: Maybe POSIX

The date and time that the revision was last updated, in ISO 8601 format.

comment :: Maybe Text

An optional comment about the revision.

httpStatus :: Int

The response's http status code.

newUpdateRevisionResponse

Arguments

:: Int
$sel:httpStatus:UpdateRevisionResponse'

-> UpdateRevisionResponse

Create a value of UpdateRevisionResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:arn:UpdateRevisionResponse', updateRevisionResponse_arn - The ARN for the revision.

$sel:createdAt:UpdateRevisionResponse', updateRevisionResponse_createdAt - The date and time that the revision was created, in ISO 8601 format.

$sel:sourceId:UpdateRevisionResponse', updateRevisionResponse_sourceId - The revision ID of the owned revision corresponding to the entitled revision being viewed. This parameter is returned when a revision owner is viewing the entitled copy of its owned revision.

$sel:finalized:UpdateRevision', updateRevisionResponse_finalized - To publish a revision to a data set in a product, the revision must first be finalized. Finalizing a revision tells AWS Data Exchange that changes to the assets in the revision are complete. After it's in this read-only state, you can publish the revision to your products.

Finalized revisions can be published through the AWS Data Exchange console or the AWS Marketplace Catalog API, using the StartChangeSet AWS Marketplace Catalog API action. When using the API, revisions are uniquely identified by their ARN.

$sel:dataSetId:UpdateRevision', updateRevisionResponse_dataSetId - The unique identifier for the data set associated with this revision.

$sel:id:UpdateRevisionResponse', updateRevisionResponse_id - The unique identifier for the revision.

$sel:updatedAt:UpdateRevisionResponse', updateRevisionResponse_updatedAt - The date and time that the revision was last updated, in ISO 8601 format.

$sel:comment:UpdateRevision', updateRevisionResponse_comment - An optional comment about the revision.

$sel:httpStatus:UpdateRevisionResponse', updateRevisionResponse_httpStatus - The response's http status code.

Response Lenses

updateRevisionResponse_arn :: Lens' UpdateRevisionResponse (Maybe Text)

The ARN for the revision.

updateRevisionResponse_createdAt :: Lens' UpdateRevisionResponse (Maybe UTCTime)

The date and time that the revision was created, in ISO 8601 format.

updateRevisionResponse_sourceId :: Lens' UpdateRevisionResponse (Maybe Text)

The revision ID of the owned revision corresponding to the entitled revision being viewed. This parameter is returned when a revision owner is viewing the entitled copy of its owned revision.

updateRevisionResponse_finalized :: Lens' UpdateRevisionResponse (Maybe Bool)

To publish a revision to a data set in a product, the revision must first be finalized. Finalizing a revision tells AWS Data Exchange that changes to the assets in the revision are complete. After it's in this read-only state, you can publish the revision to your products.

Finalized revisions can be published through the AWS Data Exchange console or the AWS Marketplace Catalog API, using the StartChangeSet AWS Marketplace Catalog API action. When using the API, revisions are uniquely identified by their ARN.

updateRevisionResponse_dataSetId :: Lens' UpdateRevisionResponse (Maybe Text)

The unique identifier for the data set associated with this revision.

updateRevisionResponse_id :: Lens' UpdateRevisionResponse (Maybe Text)

The unique identifier for the revision.

updateRevisionResponse_updatedAt :: Lens' UpdateRevisionResponse (Maybe UTCTime)

The date and time that the revision was last updated, in ISO 8601 format.

updateRevisionResponse_comment :: Lens' UpdateRevisionResponse (Maybe Text)

An optional comment about the revision.

updateRevisionResponse_httpStatus :: Lens' UpdateRevisionResponse Int

The response's http status code.