Amazonka.DataPipeline.ReportTaskRunnerHeartbeat

Description

Task runners call ReportTaskRunnerHeartbeat every 15 minutes to indicate that they are operational. If the AWS Data Pipeline Task Runner is launched on a resource managed by AWS Data Pipeline, the web service can use this call to detect when the task runner application has failed and restart a new instance.

Creating a Request

data ReportTaskRunnerHeartbeat

Contains the parameters for ReportTaskRunnerHeartbeat.

See: newReportTaskRunnerHeartbeat smart constructor.

Constructors

ReportTaskRunnerHeartbeat'

Fields

hostname :: Maybe Text

The public DNS name of the task runner.

workerGroup :: Maybe Text

The type of task the task runner is configured to accept and process. The worker group is set as a field on objects in the pipeline when they are created. You can only specify a single value for workerGroup. There are no wildcard values permitted in workerGroup; the string must be an exact, case-sensitive, match.

taskrunnerId :: Text

The ID of the task runner. This value should be unique across your AWS account. In the case of AWS Data Pipeline Task Runner launched on a resource managed by AWS Data Pipeline, the web service provides a unique identifier when it launches the application. If you have written a custom task runner, you should assign a unique identifier for the task runner.

newReportTaskRunnerHeartbeat

Arguments

:: Text
$sel:taskrunnerId:ReportTaskRunnerHeartbeat'

-> ReportTaskRunnerHeartbeat

Create a value of ReportTaskRunnerHeartbeat with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:hostname:ReportTaskRunnerHeartbeat', reportTaskRunnerHeartbeat_hostname - The public DNS name of the task runner.

$sel:workerGroup:ReportTaskRunnerHeartbeat', reportTaskRunnerHeartbeat_workerGroup - The type of task the task runner is configured to accept and process. The worker group is set as a field on objects in the pipeline when they are created. You can only specify a single value for workerGroup. There are no wildcard values permitted in workerGroup; the string must be an exact, case-sensitive, match.

$sel:taskrunnerId:ReportTaskRunnerHeartbeat', reportTaskRunnerHeartbeat_taskrunnerId - The ID of the task runner. This value should be unique across your AWS account. In the case of AWS Data Pipeline Task Runner launched on a resource managed by AWS Data Pipeline, the web service provides a unique identifier when it launches the application. If you have written a custom task runner, you should assign a unique identifier for the task runner.

Request Lenses

reportTaskRunnerHeartbeat_hostname :: Lens' ReportTaskRunnerHeartbeat (Maybe Text)

The public DNS name of the task runner.

reportTaskRunnerHeartbeat_workerGroup :: Lens' ReportTaskRunnerHeartbeat (Maybe Text)

The type of task the task runner is configured to accept and process. The worker group is set as a field on objects in the pipeline when they are created. You can only specify a single value for workerGroup. There are no wildcard values permitted in workerGroup; the string must be an exact, case-sensitive, match.

reportTaskRunnerHeartbeat_taskrunnerId :: Lens' ReportTaskRunnerHeartbeat Text

The ID of the task runner. This value should be unique across your AWS account. In the case of AWS Data Pipeline Task Runner launched on a resource managed by AWS Data Pipeline, the web service provides a unique identifier when it launches the application. If you have written a custom task runner, you should assign a unique identifier for the task runner.

Destructuring the Response

data ReportTaskRunnerHeartbeatResponse

Contains the output of ReportTaskRunnerHeartbeat.

See: newReportTaskRunnerHeartbeatResponse smart constructor.

Constructors

ReportTaskRunnerHeartbeatResponse'

Fields

httpStatus :: Int

The response's http status code.

terminate :: Bool

Indicates whether the calling task runner should terminate.

newReportTaskRunnerHeartbeatResponse

Arguments

:: Int
$sel:httpStatus:ReportTaskRunnerHeartbeatResponse'

-> Bool
$sel:terminate:ReportTaskRunnerHeartbeatResponse'

-> ReportTaskRunnerHeartbeatResponse

Create a value of ReportTaskRunnerHeartbeatResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:httpStatus:ReportTaskRunnerHeartbeatResponse', reportTaskRunnerHeartbeatResponse_httpStatus - The response's http status code.

$sel:terminate:ReportTaskRunnerHeartbeatResponse', reportTaskRunnerHeartbeatResponse_terminate - Indicates whether the calling task runner should terminate.

Response Lenses

reportTaskRunnerHeartbeatResponse_httpStatus :: Lens' ReportTaskRunnerHeartbeatResponse Int

The response's http status code.

reportTaskRunnerHeartbeatResponse_terminate :: Lens' ReportTaskRunnerHeartbeatResponse Bool

Indicates whether the calling task runner should terminate.