Amazonka.DataPipeline.Types.PipelineDescription

Description

Documentation

data PipelineDescription

Contains pipeline metadata.

See: newPipelineDescription smart constructor.

Constructors

PipelineDescription'

Fields

description :: Maybe Text

Description of the pipeline.

tags :: Maybe [Tag]

A list of tags to associated with a pipeline. Tags let you control access to pipelines. For more information, see Controlling User Access to Pipelines in the AWS Data Pipeline Developer Guide.

pipelineId :: Text

The pipeline identifier that was assigned by AWS Data Pipeline. This is a string of the form df-297EG78HU43EEXAMPLE.

name :: Text

The name of the pipeline.

fields :: [Field]

A list of read-only fields that contain metadata about the pipeline: @userId, @accountId, and @pipelineState.

newPipelineDescription

Arguments

:: Text
$sel:pipelineId:PipelineDescription'

-> Text
$sel:name:PipelineDescription'

-> PipelineDescription

Create a value of PipelineDescription with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:description:PipelineDescription', pipelineDescription_description - Description of the pipeline.

$sel:tags:PipelineDescription', pipelineDescription_tags - A list of tags to associated with a pipeline. Tags let you control access to pipelines. For more information, see Controlling User Access to Pipelines in the AWS Data Pipeline Developer Guide.

$sel:pipelineId:PipelineDescription', pipelineDescription_pipelineId - The pipeline identifier that was assigned by AWS Data Pipeline. This is a string of the form df-297EG78HU43EEXAMPLE.

$sel:name:PipelineDescription', pipelineDescription_name - The name of the pipeline.

$sel:fields:PipelineDescription', pipelineDescription_fields - A list of read-only fields that contain metadata about the pipeline: @userId, @accountId, and @pipelineState.

pipelineDescription_description :: Lens' PipelineDescription (Maybe Text)

Description of the pipeline.

pipelineDescription_tags :: Lens' PipelineDescription (Maybe [Tag])

A list of tags to associated with a pipeline. Tags let you control access to pipelines. For more information, see Controlling User Access to Pipelines in the AWS Data Pipeline Developer Guide.

pipelineDescription_pipelineId :: Lens' PipelineDescription Text

The pipeline identifier that was assigned by AWS Data Pipeline. This is a string of the form df-297EG78HU43EEXAMPLE.

pipelineDescription_name :: Lens' PipelineDescription Text

The name of the pipeline.

pipelineDescription_fields :: Lens' PipelineDescription [Field]

A list of read-only fields that contain metadata about the pipeline: @userId, @accountId, and @pipelineState.