Amazonka.DataPipeline.Types.TaskObject

Description

Documentation

data TaskObject

Contains information about a pipeline task that is assigned to a task runner.

See: newTaskObject smart constructor.

Constructors

TaskObject'

Fields

pipelineId :: Maybe Text

The ID of the pipeline that provided the task.

attemptId :: Maybe Text

The ID of the pipeline task attempt object. AWS Data Pipeline uses this value to track how many times a task is attempted.

taskId :: Maybe Text

An internal identifier for the task. This ID is passed to the SetTaskStatus and ReportTaskProgress actions.

objects :: Maybe (HashMap Text PipelineObject)

Connection information for the location where the task runner will publish the output of the task.

newTaskObject :: TaskObject

Create a value of TaskObject with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:pipelineId:TaskObject', taskObject_pipelineId - The ID of the pipeline that provided the task.

$sel:attemptId:TaskObject', taskObject_attemptId - The ID of the pipeline task attempt object. AWS Data Pipeline uses this value to track how many times a task is attempted.

$sel:taskId:TaskObject', taskObject_taskId - An internal identifier for the task. This ID is passed to the SetTaskStatus and ReportTaskProgress actions.

$sel:objects:TaskObject', taskObject_objects - Connection information for the location where the task runner will publish the output of the task.

taskObject_pipelineId :: Lens' TaskObject (Maybe Text)

The ID of the pipeline that provided the task.

taskObject_attemptId :: Lens' TaskObject (Maybe Text)

The ID of the pipeline task attempt object. AWS Data Pipeline uses this value to track how many times a task is attempted.

taskObject_taskId :: Lens' TaskObject (Maybe Text)

An internal identifier for the task. This ID is passed to the SetTaskStatus and ReportTaskProgress actions.

taskObject_objects :: Lens' TaskObject (Maybe (HashMap Text PipelineObject))

Connection information for the location where the task runner will publish the output of the task.