Amazonka.DataSync.CreateLocationEfs

Description

Creates an endpoint for an Amazon EFS file system.

Creating a Request

data CreateLocationEfs

CreateLocationEfsRequest

See: newCreateLocationEfs smart constructor.

Constructors

CreateLocationEfs'

Fields

subdirectory :: Maybe Text

A subdirectory in the location’s path. This subdirectory in the EFS file system is used to read data from the EFS source location or write data to the EFS destination. By default, DataSync uses the root directory.

Subdirectory must be specified with forward slashes. For example, /path/to/folder.

tags :: Maybe [TagListEntry]

The key-value pair that represents a tag that you want to add to the resource. The value can be an empty string. This value helps you manage, filter, and search for your resources. We recommend that you create a name tag for your location.

efsFilesystemArn :: Text

The Amazon Resource Name (ARN) for the Amazon EFS file system.

ec2Config :: Ec2Config

The subnet and security group that the Amazon EFS file system uses. The security group that you provide needs to be able to communicate with the security group on the mount target in the subnet specified.

The exact relationship between security group M (of the mount target) and security group S (which you provide for DataSync to use at this stage) is as follows:

Security group M (which you associate with the mount target) must allow inbound access for the Transmission Control Protocol (TCP) on the NFS port (2049) from security group S. You can enable inbound connections either by IP address (CIDR range) or security group.

Security group S (provided to DataSync to access EFS) should have a rule that enables outbound connections to the NFS port on one of the file system’s mount targets. You can enable outbound connections either by IP address (CIDR range) or security group.

For information about security groups and mount targets, see Security Groups for Amazon EC2 Instances and Mount Targets in the Amazon EFS User Guide.

newCreateLocationEfs

Arguments

:: Text
$sel:efsFilesystemArn:CreateLocationEfs'

-> Ec2Config
$sel:ec2Config:CreateLocationEfs'

-> CreateLocationEfs

Create a value of CreateLocationEfs with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:subdirectory:CreateLocationEfs', createLocationEfs_subdirectory - A subdirectory in the location’s path. This subdirectory in the EFS file system is used to read data from the EFS source location or write data to the EFS destination. By default, DataSync uses the root directory.

Subdirectory must be specified with forward slashes. For example, /path/to/folder.

$sel:tags:CreateLocationEfs', createLocationEfs_tags - The key-value pair that represents a tag that you want to add to the resource. The value can be an empty string. This value helps you manage, filter, and search for your resources. We recommend that you create a name tag for your location.

$sel:efsFilesystemArn:CreateLocationEfs', createLocationEfs_efsFilesystemArn - The Amazon Resource Name (ARN) for the Amazon EFS file system.

$sel:ec2Config:CreateLocationEfs', createLocationEfs_ec2Config - The subnet and security group that the Amazon EFS file system uses. The security group that you provide needs to be able to communicate with the security group on the mount target in the subnet specified.

The exact relationship between security group M (of the mount target) and security group S (which you provide for DataSync to use at this stage) is as follows:

Security group M (which you associate with the mount target) must allow inbound access for the Transmission Control Protocol (TCP) on the NFS port (2049) from security group S. You can enable inbound connections either by IP address (CIDR range) or security group.

Security group S (provided to DataSync to access EFS) should have a rule that enables outbound connections to the NFS port on one of the file system’s mount targets. You can enable outbound connections either by IP address (CIDR range) or security group.

For information about security groups and mount targets, see Security Groups for Amazon EC2 Instances and Mount Targets in the Amazon EFS User Guide.

Request Lenses

createLocationEfs_subdirectory :: Lens' CreateLocationEfs (Maybe Text)

A subdirectory in the location’s path. This subdirectory in the EFS file system is used to read data from the EFS source location or write data to the EFS destination. By default, DataSync uses the root directory.

Subdirectory must be specified with forward slashes. For example, /path/to/folder.

createLocationEfs_tags :: Lens' CreateLocationEfs (Maybe [TagListEntry])

The key-value pair that represents a tag that you want to add to the resource. The value can be an empty string. This value helps you manage, filter, and search for your resources. We recommend that you create a name tag for your location.

createLocationEfs_efsFilesystemArn :: Lens' CreateLocationEfs Text

The Amazon Resource Name (ARN) for the Amazon EFS file system.

createLocationEfs_ec2Config :: Lens' CreateLocationEfs Ec2Config

The subnet and security group that the Amazon EFS file system uses. The security group that you provide needs to be able to communicate with the security group on the mount target in the subnet specified.

The exact relationship between security group M (of the mount target) and security group S (which you provide for DataSync to use at this stage) is as follows:

Security group M (which you associate with the mount target) must allow inbound access for the Transmission Control Protocol (TCP) on the NFS port (2049) from security group S. You can enable inbound connections either by IP address (CIDR range) or security group.

Security group S (provided to DataSync to access EFS) should have a rule that enables outbound connections to the NFS port on one of the file system’s mount targets. You can enable outbound connections either by IP address (CIDR range) or security group.

For information about security groups and mount targets, see Security Groups for Amazon EC2 Instances and Mount Targets in the Amazon EFS User Guide.

Destructuring the Response

data CreateLocationEfsResponse

CreateLocationEfs

See: newCreateLocationEfsResponse smart constructor.

Constructors

CreateLocationEfsResponse'

Fields

locationArn :: Maybe Text

The Amazon Resource Name (ARN) of the Amazon EFS file system location that is created.

httpStatus :: Int

The response's http status code.

newCreateLocationEfsResponse

Arguments

:: Int
$sel:httpStatus:CreateLocationEfsResponse'

-> CreateLocationEfsResponse

Create a value of CreateLocationEfsResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:locationArn:CreateLocationEfsResponse', createLocationEfsResponse_locationArn - The Amazon Resource Name (ARN) of the Amazon EFS file system location that is created.

$sel:httpStatus:CreateLocationEfsResponse', createLocationEfsResponse_httpStatus - The response's http status code.

Response Lenses

createLocationEfsResponse_locationArn :: Lens' CreateLocationEfsResponse (Maybe Text)

The Amazon Resource Name (ARN) of the Amazon EFS file system location that is created.

createLocationEfsResponse_httpStatus :: Lens' CreateLocationEfsResponse Int

The response's http status code.