Amazonka.DataSync.ListAgents

Description

Returns a list of agents owned by an Amazon Web Services account in the Amazon Web Services Region specified in the request. The returned list is ordered by agent Amazon Resource Name (ARN).

By default, this operation returns a maximum of 100 agents. This operation supports pagination that enables you to optionally reduce the number of agents returned in a response.

If you have more agents than are returned in a response (that is, the response returns only a truncated list of your agents), the response contains a marker that you can specify in your next request to fetch the next page of agents.

This operation returns paginated results.

Creating a Request

data ListAgents

ListAgentsRequest

See: newListAgents smart constructor.

Constructors

ListAgents'

Fields

nextToken :: Maybe Text

An opaque string that indicates the position at which to begin the next list of agents.

maxResults :: Maybe Natural

The maximum number of agents to list.

newListAgents :: ListAgents

Create a value of ListAgents with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:nextToken:ListAgents', listAgents_nextToken - An opaque string that indicates the position at which to begin the next list of agents.

$sel:maxResults:ListAgents', listAgents_maxResults - The maximum number of agents to list.

Request Lenses

listAgents_nextToken :: Lens' ListAgents (Maybe Text)

An opaque string that indicates the position at which to begin the next list of agents.

listAgents_maxResults :: Lens' ListAgents (Maybe Natural)

The maximum number of agents to list.

Destructuring the Response

data ListAgentsResponse

ListAgentsResponse

See: newListAgentsResponse smart constructor.

Constructors

ListAgentsResponse'

Fields

agents :: Maybe [AgentListEntry]

A list of agents in your account.

nextToken :: Maybe Text

An opaque string that indicates the position at which to begin returning the next list of agents.

httpStatus :: Int

The response's http status code.

newListAgentsResponse

Arguments

:: Int
$sel:httpStatus:ListAgentsResponse'

-> ListAgentsResponse

Create a value of ListAgentsResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:agents:ListAgentsResponse', listAgentsResponse_agents - A list of agents in your account.

$sel:nextToken:ListAgents', listAgentsResponse_nextToken - An opaque string that indicates the position at which to begin returning the next list of agents.

$sel:httpStatus:ListAgentsResponse', listAgentsResponse_httpStatus - The response's http status code.

Response Lenses

listAgentsResponse_agents :: Lens' ListAgentsResponse (Maybe [AgentListEntry])

A list of agents in your account.

listAgentsResponse_nextToken :: Lens' ListAgentsResponse (Maybe Text)

An opaque string that indicates the position at which to begin returning the next list of agents.

listAgentsResponse_httpStatus :: Lens' ListAgentsResponse Int

The response's http status code.