Amazonka.DataSync.ListTasks

Description

Returns a list of all the tasks.

This operation returns paginated results.

Creating a Request

data ListTasks

ListTasksRequest

See: newListTasks smart constructor.

Constructors

ListTasks'

Fields

filters :: Maybe [TaskFilter]

You can use API filters to narrow down the list of resources returned by ListTasks. For example, to retrieve all tasks on a specific source location, you can use ListTasks with filter name LocationId and Operator Equals with the ARN for the location.

nextToken :: Maybe Text

An opaque string that indicates the position at which to begin the next list of tasks.

maxResults :: Maybe Natural

The maximum number of tasks to return.

newListTasks :: ListTasks

Create a value of ListTasks with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:filters:ListTasks', listTasks_filters - You can use API filters to narrow down the list of resources returned by ListTasks. For example, to retrieve all tasks on a specific source location, you can use ListTasks with filter name LocationId and Operator Equals with the ARN for the location.

$sel:nextToken:ListTasks', listTasks_nextToken - An opaque string that indicates the position at which to begin the next list of tasks.

$sel:maxResults:ListTasks', listTasks_maxResults - The maximum number of tasks to return.

Request Lenses

listTasks_filters :: Lens' ListTasks (Maybe [TaskFilter])

You can use API filters to narrow down the list of resources returned by ListTasks. For example, to retrieve all tasks on a specific source location, you can use ListTasks with filter name LocationId and Operator Equals with the ARN for the location.

listTasks_nextToken :: Lens' ListTasks (Maybe Text)

An opaque string that indicates the position at which to begin the next list of tasks.

listTasks_maxResults :: Lens' ListTasks (Maybe Natural)

The maximum number of tasks to return.

Destructuring the Response

data ListTasksResponse

ListTasksResponse

See: newListTasksResponse smart constructor.

Constructors

ListTasksResponse'

Fields

tasks :: Maybe [TaskListEntry]

A list of all the tasks that are returned.

nextToken :: Maybe Text

An opaque string that indicates the position at which to begin returning the next list of tasks.

httpStatus :: Int

The response's http status code.

newListTasksResponse

Arguments

:: Int
$sel:httpStatus:ListTasksResponse'

-> ListTasksResponse

Create a value of ListTasksResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:tasks:ListTasksResponse', listTasksResponse_tasks - A list of all the tasks that are returned.

$sel:nextToken:ListTasks', listTasksResponse_nextToken - An opaque string that indicates the position at which to begin returning the next list of tasks.

$sel:httpStatus:ListTasksResponse', listTasksResponse_httpStatus - The response's http status code.

Response Lenses

listTasksResponse_tasks :: Lens' ListTasksResponse (Maybe [TaskListEntry])

A list of all the tasks that are returned.

listTasksResponse_nextToken :: Lens' ListTasksResponse (Maybe Text)

An opaque string that indicates the position at which to begin returning the next list of tasks.

listTasksResponse_httpStatus :: Lens' ListTasksResponse Int

The response's http status code.