Amazonka.DataSync.StartTaskExecution

Description

Starts a specific invocation of a task. A TaskExecution value represents an individual run of a task. Each task can have at most one TaskExecution at a time.

TaskExecution has the following transition phases: INITIALIZING | PREPARING | TRANSFERRING | VERIFYING | SUCCESS/FAILURE.

For detailed information, see the Task Execution section in the Components and Terminology topic in the DataSync User Guide.

Creating a Request

data StartTaskExecution

StartTaskExecutionRequest

See: newStartTaskExecution smart constructor.

Constructors

StartTaskExecution'

Fields

overrideOptions :: Maybe Options

includes :: Maybe [FilterRule]

A list of filter rules that determines which files to include when running a task. The pattern should contain a single filter string that consists of the patterns to include. The patterns are delimited by "|" (that is, a pipe). For example: "/folder1|/folder2"

excludes :: Maybe [FilterRule]

A list of filter rules that determines which files to exclude from a task. The list should contain a single filter string that consists of the patterns to exclude. The patterns are delimited by "|" (that is, a pipe), for example, "/folder1|/folder2".

taskArn :: Text

The Amazon Resource Name (ARN) of the task to start.

newStartTaskExecution

Arguments

:: Text
$sel:taskArn:StartTaskExecution'

-> StartTaskExecution

Create a value of StartTaskExecution with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:overrideOptions:StartTaskExecution', startTaskExecution_overrideOptions - Undocumented member.

$sel:includes:StartTaskExecution', startTaskExecution_includes - A list of filter rules that determines which files to include when running a task. The pattern should contain a single filter string that consists of the patterns to include. The patterns are delimited by "|" (that is, a pipe). For example: "/folder1|/folder2"

$sel:excludes:StartTaskExecution', startTaskExecution_excludes - A list of filter rules that determines which files to exclude from a task. The list should contain a single filter string that consists of the patterns to exclude. The patterns are delimited by "|" (that is, a pipe), for example, "/folder1|/folder2".

$sel:taskArn:StartTaskExecution', startTaskExecution_taskArn - The Amazon Resource Name (ARN) of the task to start.

Request Lenses

startTaskExecution_overrideOptions :: Lens' StartTaskExecution (Maybe Options)

Undocumented member.

startTaskExecution_includes :: Lens' StartTaskExecution (Maybe [FilterRule])

A list of filter rules that determines which files to include when running a task. The pattern should contain a single filter string that consists of the patterns to include. The patterns are delimited by "|" (that is, a pipe). For example: "/folder1|/folder2"

startTaskExecution_excludes :: Lens' StartTaskExecution (Maybe [FilterRule])

A list of filter rules that determines which files to exclude from a task. The list should contain a single filter string that consists of the patterns to exclude. The patterns are delimited by "|" (that is, a pipe), for example, "/folder1|/folder2".

startTaskExecution_taskArn :: Lens' StartTaskExecution Text

The Amazon Resource Name (ARN) of the task to start.

Destructuring the Response

data StartTaskExecutionResponse

StartTaskExecutionResponse

See: newStartTaskExecutionResponse smart constructor.

Constructors

StartTaskExecutionResponse'

Fields

taskExecutionArn :: Maybe Text

The Amazon Resource Name (ARN) of the specific task execution that was started.

httpStatus :: Int

The response's http status code.

newStartTaskExecutionResponse

Arguments

:: Int
$sel:httpStatus:StartTaskExecutionResponse'

-> StartTaskExecutionResponse

Create a value of StartTaskExecutionResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:taskExecutionArn:StartTaskExecutionResponse', startTaskExecutionResponse_taskExecutionArn - The Amazon Resource Name (ARN) of the specific task execution that was started.

$sel:httpStatus:StartTaskExecutionResponse', startTaskExecutionResponse_httpStatus - The response's http status code.

Response Lenses

startTaskExecutionResponse_taskExecutionArn :: Lens' StartTaskExecutionResponse (Maybe Text)

The Amazon Resource Name (ARN) of the specific task execution that was started.

startTaskExecutionResponse_httpStatus :: Lens' StartTaskExecutionResponse Int

The response's http status code.