Amazonka.DataSync.UpdateLocationNfs

Description

Updates some of the parameters of a previously created location for Network File System (NFS) access. For information about creating an NFS location, see Creating a location for NFS.

Creating a Request

data UpdateLocationNfs

See: newUpdateLocationNfs smart constructor.

Constructors

UpdateLocationNfs'

Fields

onPremConfig :: Maybe OnPremConfig

subdirectory :: Maybe Text

The subdirectory in the NFS file system that is used to read data from the NFS source location or write data to the NFS destination. The NFS path should be a path that's exported by the NFS server, or a subdirectory of that path. The path should be such that it can be mounted by other NFS clients in your network.

To see all the paths exported by your NFS server, run "showmount -e nfs-server-name" from an NFS client that has access to your server. You can specify any directory that appears in the results, and any subdirectory of that directory. Ensure that the NFS export is accessible without Kerberos authentication.

To transfer all the data in the folder that you specified, DataSync must have permissions to read all the data. To ensure this, either configure the NFS export with no_root_squash, or ensure that the files you want DataSync to access have permissions that allow read access for all users. Doing either option enables the agent to read the files. For the agent to access directories, you must additionally enable all execute access.

If you are copying data to or from your Snowcone device, see NFS Server on Snowcone for more information.

For information about NFS export configuration, see 18.7. The /etc/exports Configuration File in the Red Hat Enterprise Linux documentation.

mountOptions :: Maybe NfsMountOptions

locationArn :: Text

The Amazon Resource Name (ARN) of the NFS location to update.

newUpdateLocationNfs

Arguments

:: Text
$sel:locationArn:UpdateLocationNfs'

-> UpdateLocationNfs

Create a value of UpdateLocationNfs with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:onPremConfig:UpdateLocationNfs', updateLocationNfs_onPremConfig - Undocumented member.

$sel:subdirectory:UpdateLocationNfs', updateLocationNfs_subdirectory - The subdirectory in the NFS file system that is used to read data from the NFS source location or write data to the NFS destination. The NFS path should be a path that's exported by the NFS server, or a subdirectory of that path. The path should be such that it can be mounted by other NFS clients in your network.

To see all the paths exported by your NFS server, run "showmount -e nfs-server-name" from an NFS client that has access to your server. You can specify any directory that appears in the results, and any subdirectory of that directory. Ensure that the NFS export is accessible without Kerberos authentication.

To transfer all the data in the folder that you specified, DataSync must have permissions to read all the data. To ensure this, either configure the NFS export with no_root_squash, or ensure that the files you want DataSync to access have permissions that allow read access for all users. Doing either option enables the agent to read the files. For the agent to access directories, you must additionally enable all execute access.

If you are copying data to or from your Snowcone device, see NFS Server on Snowcone for more information.

For information about NFS export configuration, see 18.7. The /etc/exports Configuration File in the Red Hat Enterprise Linux documentation.

$sel:mountOptions:UpdateLocationNfs', updateLocationNfs_mountOptions - Undocumented member.

$sel:locationArn:UpdateLocationNfs', updateLocationNfs_locationArn - The Amazon Resource Name (ARN) of the NFS location to update.

Request Lenses

updateLocationNfs_onPremConfig :: Lens' UpdateLocationNfs (Maybe OnPremConfig)

Undocumented member.

updateLocationNfs_subdirectory :: Lens' UpdateLocationNfs (Maybe Text)

The subdirectory in the NFS file system that is used to read data from the NFS source location or write data to the NFS destination. The NFS path should be a path that's exported by the NFS server, or a subdirectory of that path. The path should be such that it can be mounted by other NFS clients in your network.

To see all the paths exported by your NFS server, run "showmount -e nfs-server-name" from an NFS client that has access to your server. You can specify any directory that appears in the results, and any subdirectory of that directory. Ensure that the NFS export is accessible without Kerberos authentication.

To transfer all the data in the folder that you specified, DataSync must have permissions to read all the data. To ensure this, either configure the NFS export with no_root_squash, or ensure that the files you want DataSync to access have permissions that allow read access for all users. Doing either option enables the agent to read the files. For the agent to access directories, you must additionally enable all execute access.

If you are copying data to or from your Snowcone device, see NFS Server on Snowcone for more information.

For information about NFS export configuration, see 18.7. The /etc/exports Configuration File in the Red Hat Enterprise Linux documentation.

updateLocationNfs_mountOptions :: Lens' UpdateLocationNfs (Maybe NfsMountOptions)

Undocumented member.

updateLocationNfs_locationArn :: Lens' UpdateLocationNfs Text

The Amazon Resource Name (ARN) of the NFS location to update.

Destructuring the Response

data UpdateLocationNfsResponse

See: newUpdateLocationNfsResponse smart constructor.

Constructors

UpdateLocationNfsResponse'

Fields

httpStatus :: Int

The response's http status code.

newUpdateLocationNfsResponse

Arguments

:: Int
$sel:httpStatus:UpdateLocationNfsResponse'

-> UpdateLocationNfsResponse

Create a value of UpdateLocationNfsResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:httpStatus:UpdateLocationNfsResponse', updateLocationNfsResponse_httpStatus - The response's http status code.

Response Lenses

updateLocationNfsResponse_httpStatus :: Lens' UpdateLocationNfsResponse Int

The response's http status code.