Amazonka.DataSync.UpdateTask

Description

Updates the metadata associated with a task.

Creating a Request

data UpdateTask

UpdateTaskResponse

See: newUpdateTask smart constructor.

Constructors

UpdateTask'

Fields

schedule :: Maybe TaskSchedule

Specifies a schedule used to periodically transfer files from a source to a destination location. You can configure your task to execute hourly, daily, weekly or on specific days of the week. You control when in the day or hour you want the task to execute. The time you specify is UTC time. For more information, see Scheduling your task.

includes :: Maybe [FilterRule]

A list of filter rules that determines which files to include when running a task. The pattern should contain a single filter string that consists of the patterns to include. The patterns are delimited by "|" (that is, a pipe). For example: "/folder1|/folder2"

name :: Maybe Text

The name of the task to update.

excludes :: Maybe [FilterRule]

A list of filter rules that determines which files to exclude from a task. The list should contain a single filter string that consists of the patterns to exclude. The patterns are delimited by "|" (that is, a pipe), for example: "/folder1|/folder2"

options :: Maybe Options

cloudWatchLogGroupArn :: Maybe Text

The Amazon Resource Name (ARN) of the resource name of the CloudWatch LogGroup.

taskArn :: Text

The Amazon Resource Name (ARN) of the resource name of the task to update.

newUpdateTask

Arguments

:: Text
$sel:taskArn:UpdateTask'

-> UpdateTask

Create a value of UpdateTask with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:schedule:UpdateTask', updateTask_schedule - Specifies a schedule used to periodically transfer files from a source to a destination location. You can configure your task to execute hourly, daily, weekly or on specific days of the week. You control when in the day or hour you want the task to execute. The time you specify is UTC time. For more information, see Scheduling your task.

$sel:includes:UpdateTask', updateTask_includes - A list of filter rules that determines which files to include when running a task. The pattern should contain a single filter string that consists of the patterns to include. The patterns are delimited by "|" (that is, a pipe). For example: "/folder1|/folder2"

$sel:name:UpdateTask', updateTask_name - The name of the task to update.

$sel:excludes:UpdateTask', updateTask_excludes - A list of filter rules that determines which files to exclude from a task. The list should contain a single filter string that consists of the patterns to exclude. The patterns are delimited by "|" (that is, a pipe), for example: "/folder1|/folder2"

$sel:options:UpdateTask', updateTask_options - Undocumented member.

$sel:cloudWatchLogGroupArn:UpdateTask', updateTask_cloudWatchLogGroupArn - The Amazon Resource Name (ARN) of the resource name of the CloudWatch LogGroup.

$sel:taskArn:UpdateTask', updateTask_taskArn - The Amazon Resource Name (ARN) of the resource name of the task to update.

Request Lenses

updateTask_schedule :: Lens' UpdateTask (Maybe TaskSchedule)

Specifies a schedule used to periodically transfer files from a source to a destination location. You can configure your task to execute hourly, daily, weekly or on specific days of the week. You control when in the day or hour you want the task to execute. The time you specify is UTC time. For more information, see Scheduling your task.

updateTask_includes :: Lens' UpdateTask (Maybe [FilterRule])

A list of filter rules that determines which files to include when running a task. The pattern should contain a single filter string that consists of the patterns to include. The patterns are delimited by "|" (that is, a pipe). For example: "/folder1|/folder2"

updateTask_name :: Lens' UpdateTask (Maybe Text)

The name of the task to update.

updateTask_excludes :: Lens' UpdateTask (Maybe [FilterRule])

A list of filter rules that determines which files to exclude from a task. The list should contain a single filter string that consists of the patterns to exclude. The patterns are delimited by "|" (that is, a pipe), for example: "/folder1|/folder2"

updateTask_options :: Lens' UpdateTask (Maybe Options)

Undocumented member.

updateTask_cloudWatchLogGroupArn :: Lens' UpdateTask (Maybe Text)

The Amazon Resource Name (ARN) of the resource name of the CloudWatch LogGroup.

updateTask_taskArn :: Lens' UpdateTask Text

The Amazon Resource Name (ARN) of the resource name of the task to update.

Destructuring the Response

data UpdateTaskResponse

See: newUpdateTaskResponse smart constructor.

Constructors

UpdateTaskResponse'

Fields

httpStatus :: Int

The response's http status code.

newUpdateTaskResponse

Arguments

:: Int
$sel:httpStatus:UpdateTaskResponse'

-> UpdateTaskResponse

Create a value of UpdateTaskResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:httpStatus:UpdateTaskResponse', updateTaskResponse_httpStatus - The response's http status code.

Response Lenses

updateTaskResponse_httpStatus :: Lens' UpdateTaskResponse Int

The response's http status code.