Amazonka.Detective

Description

Derived from API version 2018-10-26 of the AWS service descriptions, licensed under Apache 2.0.

Detective uses machine learning and purpose-built visualizations to help you analyze and investigate security issues across your Amazon Web Services (AWS) workloads. Detective automatically extracts time-based events such as login attempts, API calls, and network traffic from AWS CloudTrail and Amazon Virtual Private Cloud (Amazon VPC) flow logs. It also extracts findings detected by Amazon GuardDuty.

The Detective API primarily supports the creation and management of behavior graphs. A behavior graph contains the extracted data from a set of member accounts, and is created and managed by an administrator account.

Every behavior graph is specific to a Region. You can only use the API to manage graphs that belong to the Region that is associated with the currently selected endpoint.

A Detective administrator account can use the Detective API to do the following:

Enable and disable Detective. Enabling Detective creates a new behavior graph.

View the list of member accounts in a behavior graph.

Add member accounts to a behavior graph.

Remove member accounts from a behavior graph.

A member account can use the Detective API to do the following:

View the list of behavior graphs that they are invited to.

Accept an invitation to contribute to a behavior graph.

Decline an invitation to contribute to a behavior graph.

Remove their account from a behavior graph.

All API actions are logged as CloudTrail events. See Logging Detective API Calls with CloudTrail.

We replaced the term "master account" with the term "administrator account." An administrator account is used to centrally manage multiple accounts. In the case of Detective, the administrator account manages the accounts in their behavior graph.

Service Configuration

defaultService :: Service

API version 2018-10-26 of the Amazon Detective SDK configuration.

Errors

Error matchers are designed for use with the functions provided by Control.Exception.Lens. This allows catching (and rethrowing) service specific errors returned by Detective.

ValidationException

_ValidationException :: AsError a => Getting (First ServiceError) a ServiceError

The request parameters are invalid.

ConflictException

_ConflictException :: AsError a => Getting (First ServiceError) a ServiceError

The request attempted an invalid action.

ServiceQuotaExceededException

_ServiceQuotaExceededException :: AsError a => Getting (First ServiceError) a ServiceError

This request cannot be completed for one of the following reasons.

The request would cause the number of member accounts in the behavior graph to exceed the maximum allowed. A behavior graph cannot have more than 1000 member accounts.

The request would cause the data rate for the behavior graph to exceed the maximum allowed.

Detective is unable to verify the data rate for the member account. This is usually because the member account is not enrolled in Amazon GuardDuty.

InternalServerException

_InternalServerException :: AsError a => Getting (First ServiceError) a ServiceError

The request was valid but failed because of a problem with the service.

ResourceNotFoundException

_ResourceNotFoundException :: AsError a => Getting (First ServiceError) a ServiceError

The request refers to a nonexistent resource.

Waiters

Waiters poll by repeatedly sending a request until some remote success condition configured by the Wait specification is fulfilled. The Wait specification determines how many attempts should be made, in addition to delay and retry strategies.

Operations

Some AWS operations return results that are incomplete and require subsequent requests in order to obtain the entire result set. The process of sending subsequent requests to continue where a previous request left off is called pagination. For example, the ListObjects operation of Amazon S3 returns up to 1000 objects at a time, and you must send subsequent requests with the appropriate Marker in order to retrieve the next page of results.

Operations that have an AWSPager instance can transparently perform subsequent requests, correctly setting Markers and other request facets to iterate through the entire result set of a truncated API operation. Operations which support this have an additional note in the documentation.

Many operations have the ability to filter results on the server side. See the individual operation parameters for details.

StartMonitoringMember

data StartMonitoringMember

See: newStartMonitoringMember smart constructor.

Constructors

StartMonitoringMember' Text Text

newStartMonitoringMember

Arguments

:: Text
$sel:graphArn:StartMonitoringMember'

-> Text
$sel:accountId:StartMonitoringMember'

-> StartMonitoringMember

Create a value of StartMonitoringMember with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:graphArn:StartMonitoringMember', startMonitoringMember_graphArn - The ARN of the behavior graph.

$sel:accountId:StartMonitoringMember', startMonitoringMember_accountId - The account ID of the member account to try to enable.

The account must be an invited member account with a status of ACCEPTED_BUT_DISABLED.

data StartMonitoringMemberResponse

See: newStartMonitoringMemberResponse smart constructor.

Constructors

StartMonitoringMemberResponse'

newStartMonitoringMemberResponse :: StartMonitoringMemberResponse

Create a value of StartMonitoringMemberResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

DeleteMembers

data DeleteMembers

See: newDeleteMembers smart constructor.

Constructors

DeleteMembers' Text (NonEmpty Text)

newDeleteMembers

Arguments

:: Text
$sel:graphArn:DeleteMembers'

-> NonEmpty Text
$sel:accountIds:DeleteMembers'

-> DeleteMembers

Create a value of DeleteMembers with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:graphArn:DeleteMembers', deleteMembers_graphArn - The ARN of the behavior graph to delete members from.

$sel:accountIds:DeleteMembers', deleteMembers_accountIds - The list of AWS account identifiers for the member accounts to delete from the behavior graph. You can delete up to 50 member accounts at a time.

data DeleteMembersResponse

See: newDeleteMembersResponse smart constructor.

Constructors

DeleteMembersResponse' (Maybe (NonEmpty Text)) (Maybe [UnprocessedAccount]) Int

newDeleteMembersResponse

Arguments

:: Int
$sel:httpStatus:DeleteMembersResponse'

-> DeleteMembersResponse

Create a value of DeleteMembersResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:accountIds:DeleteMembers', deleteMembersResponse_accountIds - The list of AWS account identifiers for the member accounts that Detective successfully deleted from the behavior graph.

$sel:unprocessedAccounts:DeleteMembersResponse', deleteMembersResponse_unprocessedAccounts - The list of member accounts that Detective was not able to delete from the behavior graph. For each member account, provides the reason that the deletion could not be processed.

$sel:httpStatus:DeleteMembersResponse', deleteMembersResponse_httpStatus - The response's http status code.

ListTagsForResource

data ListTagsForResource

See: newListTagsForResource smart constructor.

Constructors

ListTagsForResource' Text

newListTagsForResource

Arguments

:: Text
$sel:resourceArn:ListTagsForResource'

-> ListTagsForResource

Create a value of ListTagsForResource with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:resourceArn:ListTagsForResource', listTagsForResource_resourceArn - The ARN of the behavior graph for which to retrieve the tag values.

data ListTagsForResourceResponse

See: newListTagsForResourceResponse smart constructor.

Constructors

ListTagsForResourceResponse' (Maybe (HashMap Text Text)) Int

newListTagsForResourceResponse

Arguments

:: Int
$sel:httpStatus:ListTagsForResourceResponse'

-> ListTagsForResourceResponse

Create a value of ListTagsForResourceResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:tags:ListTagsForResourceResponse', listTagsForResourceResponse_tags - The tag values that are assigned to the behavior graph. The request returns up to 50 tag values.

$sel:httpStatus:ListTagsForResourceResponse', listTagsForResourceResponse_httpStatus - The response's http status code.

DeleteGraph

data DeleteGraph

See: newDeleteGraph smart constructor.

Constructors

DeleteGraph' Text

newDeleteGraph

Arguments

:: Text
$sel:graphArn:DeleteGraph'

-> DeleteGraph

Create a value of DeleteGraph with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:graphArn:DeleteGraph', deleteGraph_graphArn - The ARN of the behavior graph to disable.

data DeleteGraphResponse

See: newDeleteGraphResponse smart constructor.

Constructors

DeleteGraphResponse'

newDeleteGraphResponse :: DeleteGraphResponse

Create a value of DeleteGraphResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

ListInvitations

data ListInvitations

See: newListInvitations smart constructor.

Constructors

ListInvitations' (Maybe Text) (Maybe Natural)

newListInvitations :: ListInvitations

Create a value of ListInvitations with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:nextToken:ListInvitations', listInvitations_nextToken - For requests to retrieve the next page of results, the pagination token that was returned with the previous page of results. The initial request does not include a pagination token.

$sel:maxResults:ListInvitations', listInvitations_maxResults - The maximum number of behavior graph invitations to return in the response. The total must be less than the overall limit on the number of results to return, which is currently 200.

data ListInvitationsResponse

See: newListInvitationsResponse smart constructor.

Constructors

ListInvitationsResponse' (Maybe [MemberDetail]) (Maybe Text) Int

newListInvitationsResponse

Arguments

:: Int
$sel:httpStatus:ListInvitationsResponse'

-> ListInvitationsResponse

Create a value of ListInvitationsResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:invitations:ListInvitationsResponse', listInvitationsResponse_invitations - The list of behavior graphs for which the member account has open or accepted invitations.

$sel:nextToken:ListInvitations', listInvitationsResponse_nextToken - If there are more behavior graphs remaining in the results, then this is the pagination token to use to request the next page of behavior graphs.

$sel:httpStatus:ListInvitationsResponse', listInvitationsResponse_httpStatus - The response's http status code.

DisassociateMembership

data DisassociateMembership

See: newDisassociateMembership smart constructor.

Constructors

DisassociateMembership' Text

newDisassociateMembership

Arguments

:: Text
$sel:graphArn:DisassociateMembership'

-> DisassociateMembership

Create a value of DisassociateMembership with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:graphArn:DisassociateMembership', disassociateMembership_graphArn - The ARN of the behavior graph to remove the member account from.

The member account's member status in the behavior graph must be ENABLED.

data DisassociateMembershipResponse

See: newDisassociateMembershipResponse smart constructor.

Constructors

DisassociateMembershipResponse'

newDisassociateMembershipResponse :: DisassociateMembershipResponse

Create a value of DisassociateMembershipResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

AcceptInvitation

data AcceptInvitation

See: newAcceptInvitation smart constructor.

Constructors

AcceptInvitation' Text

newAcceptInvitation

Arguments

:: Text
$sel:graphArn:AcceptInvitation'

-> AcceptInvitation

Create a value of AcceptInvitation with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:graphArn:AcceptInvitation', acceptInvitation_graphArn - The ARN of the behavior graph that the member account is accepting the invitation for.

The member account status in the behavior graph must be INVITED.

data AcceptInvitationResponse

See: newAcceptInvitationResponse smart constructor.

Constructors

AcceptInvitationResponse'

newAcceptInvitationResponse :: AcceptInvitationResponse

Create a value of AcceptInvitationResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

ListMembers

data ListMembers

See: newListMembers smart constructor.

Constructors

ListMembers' (Maybe Text) (Maybe Natural) Text

newListMembers

Arguments

:: Text
$sel:graphArn:ListMembers'

-> ListMembers

Create a value of ListMembers with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:nextToken:ListMembers', listMembers_nextToken - For requests to retrieve the next page of member account results, the pagination token that was returned with the previous page of results. The initial request does not include a pagination token.

$sel:maxResults:ListMembers', listMembers_maxResults - The maximum number of member accounts to include in the response. The total must be less than the overall limit on the number of results to return, which is currently 200.

$sel:graphArn:ListMembers', listMembers_graphArn - The ARN of the behavior graph for which to retrieve the list of member accounts.

data ListMembersResponse

See: newListMembersResponse smart constructor.

Constructors

ListMembersResponse' (Maybe [MemberDetail]) (Maybe Text) Int

newListMembersResponse

Arguments

:: Int
$sel:httpStatus:ListMembersResponse'

-> ListMembersResponse

Create a value of ListMembersResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:memberDetails:ListMembersResponse', listMembersResponse_memberDetails - The list of member accounts in the behavior graph.

The results include member accounts that did not pass verification and member accounts that have not yet accepted the invitation to the behavior graph. The results do not include member accounts that were removed from the behavior graph.

$sel:nextToken:ListMembers', listMembersResponse_nextToken - If there are more member accounts remaining in the results, then this is the pagination token to use to request the next page of member accounts.

$sel:httpStatus:ListMembersResponse', listMembersResponse_httpStatus - The response's http status code.

CreateMembers

data CreateMembers

See: newCreateMembers smart constructor.

Constructors

CreateMembers' (Maybe Bool) (Maybe Text) Text (NonEmpty Account)

newCreateMembers

Arguments

:: Text
$sel:graphArn:CreateMembers'

-> NonEmpty Account
$sel:accounts:CreateMembers'

-> CreateMembers

Create a value of CreateMembers with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:disableEmailNotification:CreateMembers', createMembers_disableEmailNotification - if set to true, then the member accounts do not receive email notifications. By default, this is set to false, and the member accounts receive email notifications.

$sel:message:CreateMembers', createMembers_message - Customized message text to include in the invitation email message to the invited member accounts.

$sel:graphArn:CreateMembers', createMembers_graphArn - The ARN of the behavior graph to invite the member accounts to contribute their data to.

$sel:accounts:CreateMembers', createMembers_accounts - The list of AWS accounts to invite to become member accounts in the behavior graph. You can invite up to 50 accounts at a time. For each invited account, the account list contains the account identifier and the AWS account root user email address.

data CreateMembersResponse

See: newCreateMembersResponse smart constructor.

Constructors

CreateMembersResponse' (Maybe [MemberDetail]) (Maybe [UnprocessedAccount]) Int

newCreateMembersResponse

Arguments

:: Int
$sel:httpStatus:CreateMembersResponse'

-> CreateMembersResponse

Create a value of CreateMembersResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:members:CreateMembersResponse', createMembersResponse_members - The set of member account invitation requests that Detective was able to process. This includes accounts that are being verified, that failed verification, and that passed verification and are being sent an invitation.

$sel:unprocessedAccounts:CreateMembersResponse', createMembersResponse_unprocessedAccounts - The list of accounts for which Detective was unable to process the invitation request. For each account, the list provides the reason why the request could not be processed. The list includes accounts that are already member accounts in the behavior graph.

$sel:httpStatus:CreateMembersResponse', createMembersResponse_httpStatus - The response's http status code.

GetMembers

data GetMembers

See: newGetMembers smart constructor.

Constructors

GetMembers' Text (NonEmpty Text)

newGetMembers

Arguments

:: Text
$sel:graphArn:GetMembers'

-> NonEmpty Text
$sel:accountIds:GetMembers'

-> GetMembers

Create a value of GetMembers with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:graphArn:GetMembers', getMembers_graphArn - The ARN of the behavior graph for which to request the member details.

$sel:accountIds:GetMembers', getMembers_accountIds - The list of AWS account identifiers for the member account for which to return member details. You can request details for up to 50 member accounts at a time.

You cannot use GetMembers to retrieve information about member accounts that were removed from the behavior graph.

data GetMembersResponse

See: newGetMembersResponse smart constructor.

Constructors

GetMembersResponse' (Maybe [MemberDetail]) (Maybe [UnprocessedAccount]) Int

newGetMembersResponse

Arguments

:: Int
$sel:httpStatus:GetMembersResponse'

-> GetMembersResponse

Create a value of GetMembersResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:memberDetails:GetMembersResponse', getMembersResponse_memberDetails - The member account details that Detective is returning in response to the request.

$sel:unprocessedAccounts:GetMembersResponse', getMembersResponse_unprocessedAccounts - The requested member accounts for which Detective was unable to return member details.

For each account, provides the reason why the request could not be processed.

$sel:httpStatus:GetMembersResponse', getMembersResponse_httpStatus - The response's http status code.

ListGraphs

data ListGraphs

See: newListGraphs smart constructor.

Constructors

ListGraphs' (Maybe Text) (Maybe Natural)

newListGraphs :: ListGraphs

Create a value of ListGraphs with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:nextToken:ListGraphs', listGraphs_nextToken - For requests to get the next page of results, the pagination token that was returned with the previous set of results. The initial request does not include a pagination token.

$sel:maxResults:ListGraphs', listGraphs_maxResults - The maximum number of graphs to return at a time. The total must be less than the overall limit on the number of results to return, which is currently 200.

data ListGraphsResponse

See: newListGraphsResponse smart constructor.

Constructors

ListGraphsResponse' (Maybe Text) (Maybe [Graph]) Int

newListGraphsResponse

Arguments

:: Int
$sel:httpStatus:ListGraphsResponse'

-> ListGraphsResponse

Create a value of ListGraphsResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:nextToken:ListGraphs', listGraphsResponse_nextToken - If there are more behavior graphs remaining in the results, then this is the pagination token to use to request the next page of behavior graphs.

$sel:graphList:ListGraphsResponse', listGraphsResponse_graphList - A list of behavior graphs that the account is an administrator account for.

$sel:httpStatus:ListGraphsResponse', listGraphsResponse_httpStatus - The response's http status code.

TagResource

data TagResource

See: newTagResource smart constructor.

Constructors

TagResource' Text (HashMap Text Text)

newTagResource

Arguments

:: Text
$sel:resourceArn:TagResource'

-> TagResource

Create a value of TagResource with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:resourceArn:TagResource', tagResource_resourceArn - The ARN of the behavior graph to assign the tags to.

$sel:tags:TagResource', tagResource_tags - The tags to assign to the behavior graph. You can add up to 50 tags. For each tag, you provide the tag key and the tag value. Each tag key can contain up to 128 characters. Each tag value can contain up to 256 characters.

data TagResourceResponse

See: newTagResourceResponse smart constructor.

Constructors

TagResourceResponse' Int

newTagResourceResponse

Arguments

:: Int
$sel:httpStatus:TagResourceResponse'

-> TagResourceResponse

Create a value of TagResourceResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:httpStatus:TagResourceResponse', tagResourceResponse_httpStatus - The response's http status code.

CreateGraph

data CreateGraph

See: newCreateGraph smart constructor.

Constructors

CreateGraph' (Maybe (HashMap Text Text))

newCreateGraph :: CreateGraph

Create a value of CreateGraph with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:tags:CreateGraph', createGraph_tags - The tags to assign to the new behavior graph. You can add up to 50 tags. For each tag, you provide the tag key and the tag value. Each tag key can contain up to 128 characters. Each tag value can contain up to 256 characters.

data CreateGraphResponse

See: newCreateGraphResponse smart constructor.

Constructors

CreateGraphResponse' (Maybe Text) Int

newCreateGraphResponse

Arguments

:: Int
$sel:httpStatus:CreateGraphResponse'

-> CreateGraphResponse

Create a value of CreateGraphResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:graphArn:CreateGraphResponse', createGraphResponse_graphArn - The ARN of the new behavior graph.

$sel:httpStatus:CreateGraphResponse', createGraphResponse_httpStatus - The response's http status code.

UntagResource

data UntagResource

See: newUntagResource smart constructor.

Constructors

UntagResource' Text (NonEmpty Text)

newUntagResource

Arguments

:: Text
$sel:resourceArn:UntagResource'

-> NonEmpty Text
$sel:tagKeys:UntagResource'

-> UntagResource

Create a value of UntagResource with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:resourceArn:UntagResource', untagResource_resourceArn - The ARN of the behavior graph to remove the tags from.

$sel:tagKeys:UntagResource', untagResource_tagKeys - The tag keys of the tags to remove from the behavior graph. You can remove up to 50 tags at a time.

data UntagResourceResponse

See: newUntagResourceResponse smart constructor.

Constructors

UntagResourceResponse' Int

newUntagResourceResponse

Arguments

:: Int
$sel:httpStatus:UntagResourceResponse'

-> UntagResourceResponse

Create a value of UntagResourceResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:httpStatus:UntagResourceResponse', untagResourceResponse_httpStatus - The response's http status code.

RejectInvitation

data RejectInvitation

See: newRejectInvitation smart constructor.

Constructors

RejectInvitation' Text

newRejectInvitation

Arguments

:: Text
$sel:graphArn:RejectInvitation'

-> RejectInvitation

Create a value of RejectInvitation with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:graphArn:RejectInvitation', rejectInvitation_graphArn - The ARN of the behavior graph to reject the invitation to.

The member account's current member status in the behavior graph must be INVITED.

data RejectInvitationResponse

See: newRejectInvitationResponse smart constructor.

Constructors

RejectInvitationResponse'

newRejectInvitationResponse :: RejectInvitationResponse

Create a value of RejectInvitationResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

Types

MemberDisabledReason

newtype MemberDisabledReason

Constructors

MemberDisabledReason'

Fields

fromMemberDisabledReason :: Text

Bundled Patterns

pattern MemberDisabledReason_VOLUME_TOO_HIGH :: MemberDisabledReason

pattern MemberDisabledReason_VOLUME_UNKNOWN :: MemberDisabledReason

MemberStatus

newtype MemberStatus

Constructors

MemberStatus'

Fields

fromMemberStatus :: Text

Bundled Patterns

pattern MemberStatus_ACCEPTED_BUT_DISABLED :: MemberStatus

pattern MemberStatus_ENABLED :: MemberStatus

pattern MemberStatus_INVITED :: MemberStatus

pattern MemberStatus_VERIFICATION_FAILED :: MemberStatus

pattern MemberStatus_VERIFICATION_IN_PROGRESS :: MemberStatus

Account

data Account

An AWS account that is the administrator account of or a member of a behavior graph.

See: newAccount smart constructor.

Constructors

Account' Text Text

newAccount

Arguments

:: Text
$sel:accountId:Account'

-> Text
$sel:emailAddress:Account'

-> Account

Create a value of Account with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:accountId:Account', account_accountId - The account identifier of the AWS account.

$sel:emailAddress:Account', account_emailAddress - The AWS account root user email address for the AWS account.

Graph

data Graph

A behavior graph in Detective.

See: newGraph smart constructor.

Constructors

Graph' (Maybe Text) (Maybe POSIX)

newGraph :: Graph

Create a value of Graph with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:arn:Graph', graph_arn - The ARN of the behavior graph.

$sel:createdTime:Graph', graph_createdTime - The date and time that the behavior graph was created. The value is in milliseconds since the epoch.

MemberDetail

data MemberDetail

Details about a member account that was invited to contribute to a behavior graph.

See: newMemberDetail smart constructor.

Constructors

MemberDetail' (Maybe POSIX) (Maybe MemberStatus) (Maybe POSIX) (Maybe Text) (Maybe Text) (Maybe Text) (Maybe Text) (Maybe MemberDisabledReason) (Maybe Double) (Maybe Text) (Maybe POSIX) (Maybe POSIX) (Maybe Integer)

newMemberDetail :: MemberDetail

Create a value of MemberDetail with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:percentOfGraphUtilizationUpdatedTime:MemberDetail', memberDetail_percentOfGraphUtilizationUpdatedTime - The date and time when the graph utilization percentage was last updated.

$sel:status:MemberDetail', memberDetail_status - The current membership status of the member account. The status can have one of the following values:

INVITED - Indicates that the member was sent an invitation but has not yet responded.

VERIFICATION_IN_PROGRESS - Indicates that Detective is verifying that the account identifier and email address provided for the member account match. If they do match, then Detective sends the invitation. If the email address and account identifier don't match, then the member cannot be added to the behavior graph.

VERIFICATION_FAILED - Indicates that the account and email address provided for the member account do not match, and Detective did not send an invitation to the account.

ENABLED - Indicates that the member account accepted the invitation to contribute to the behavior graph.

ACCEPTED_BUT_DISABLED - Indicates that the member account accepted the invitation but is prevented from contributing data to the behavior graph. DisabledReason provides the reason why the member account is not enabled.

Member accounts that declined an invitation or that were removed from the behavior graph are not included.

$sel:invitedTime:MemberDetail', memberDetail_invitedTime - The date and time that Detective sent the invitation to the member account. The value is in milliseconds since the epoch.

$sel:administratorId:MemberDetail', memberDetail_administratorId - The AWS account identifier of the administrator account for the behavior graph.

$sel:graphArn:MemberDetail', memberDetail_graphArn - The ARN of the behavior graph that the member account was invited to.

$sel:masterId:MemberDetail', memberDetail_masterId - The AWS account identifier of the administrator account for the behavior graph.

$sel:accountId:MemberDetail', memberDetail_accountId - The AWS account identifier for the member account.

$sel:disabledReason:MemberDetail', memberDetail_disabledReason - For member accounts with a status of ACCEPTED_BUT_DISABLED, the reason that the member account is not enabled.

The reason can have one of the following values:

VOLUME_TOO_HIGH - Indicates that adding the member account would cause the data volume for the behavior graph to be too high.

VOLUME_UNKNOWN - Indicates that Detective is unable to verify the data volume for the member account. This is usually because the member account is not enrolled in Amazon GuardDuty.

$sel:percentOfGraphUtilization:MemberDetail', memberDetail_percentOfGraphUtilization - The member account data volume as a percentage of the maximum allowed data volume. 0 indicates 0 percent, and 100 indicates 100 percent.

Note that this is not the percentage of the behavior graph data volume.

For example, the data volume for the behavior graph is 80 GB per day. The maximum data volume is 160 GB per day. If the data volume for the member account is 40 GB per day, then PercentOfGraphUtilization is 25. It represents 25% of the maximum allowed data volume.

$sel:emailAddress:MemberDetail', memberDetail_emailAddress - The AWS account root user email address for the member account.

$sel:volumeUsageUpdatedTime:MemberDetail', memberDetail_volumeUsageUpdatedTime - The data and time when the member account data volume was last updated.

$sel:updatedTime:MemberDetail', memberDetail_updatedTime - The date and time that the member account was last updated. The value is in milliseconds since the epoch.

$sel:volumeUsageInBytes:MemberDetail', memberDetail_volumeUsageInBytes - The data volume in bytes per day for the member account.

UnprocessedAccount

data UnprocessedAccount

A member account that was included in a request but for which the request could not be processed.

See: newUnprocessedAccount smart constructor.

Constructors

UnprocessedAccount' (Maybe Text) (Maybe Text)

newUnprocessedAccount :: UnprocessedAccount

Create a value of UnprocessedAccount with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:accountId:UnprocessedAccount', unprocessedAccount_accountId - The AWS account identifier of the member account that was not processed.

$sel:reason:UnprocessedAccount', unprocessedAccount_reason - The reason that the member account request could not be processed.