Amazonka.Detective.AcceptInvitation

Description

Accepts an invitation for the member account to contribute data to a behavior graph. This operation can only be called by an invited member account.

The request provides the ARN of behavior graph.

The member account status in the graph must be INVITED.

Creating a Request

data AcceptInvitation

See: newAcceptInvitation smart constructor.

Constructors

AcceptInvitation'

Fields

graphArn :: Text

The ARN of the behavior graph that the member account is accepting the invitation for.

The member account status in the behavior graph must be INVITED.

newAcceptInvitation

Arguments

:: Text
$sel:graphArn:AcceptInvitation'

-> AcceptInvitation

Create a value of AcceptInvitation with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:graphArn:AcceptInvitation', acceptInvitation_graphArn - The ARN of the behavior graph that the member account is accepting the invitation for.

The member account status in the behavior graph must be INVITED.

Request Lenses

acceptInvitation_graphArn :: Lens' AcceptInvitation Text

The ARN of the behavior graph that the member account is accepting the invitation for.

The member account status in the behavior graph must be INVITED.

Destructuring the Response

data AcceptInvitationResponse

See: newAcceptInvitationResponse smart constructor.

Constructors

AcceptInvitationResponse'

newAcceptInvitationResponse :: AcceptInvitationResponse

Create a value of AcceptInvitationResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.