Amazonka.Detective.GetMembers

Description

Returns the membership details for specified member accounts for a behavior graph.

Creating a Request

data GetMembers

See: newGetMembers smart constructor.

Constructors

GetMembers'

Fields

graphArn :: Text

The ARN of the behavior graph for which to request the member details.

accountIds :: NonEmpty Text

The list of AWS account identifiers for the member account for which to return member details. You can request details for up to 50 member accounts at a time.

You cannot use GetMembers to retrieve information about member accounts that were removed from the behavior graph.

newGetMembers

Arguments

:: Text
$sel:graphArn:GetMembers'

-> NonEmpty Text
$sel:accountIds:GetMembers'

-> GetMembers

Create a value of GetMembers with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:graphArn:GetMembers', getMembers_graphArn - The ARN of the behavior graph for which to request the member details.

$sel:accountIds:GetMembers', getMembers_accountIds - The list of AWS account identifiers for the member account for which to return member details. You can request details for up to 50 member accounts at a time.

You cannot use GetMembers to retrieve information about member accounts that were removed from the behavior graph.

Request Lenses

getMembers_graphArn :: Lens' GetMembers Text

The ARN of the behavior graph for which to request the member details.

getMembers_accountIds :: Lens' GetMembers (NonEmpty Text)

The list of AWS account identifiers for the member account for which to return member details. You can request details for up to 50 member accounts at a time.

You cannot use GetMembers to retrieve information about member accounts that were removed from the behavior graph.

Destructuring the Response

data GetMembersResponse

See: newGetMembersResponse smart constructor.

Constructors

GetMembersResponse'

Fields

memberDetails :: Maybe [MemberDetail]

The member account details that Detective is returning in response to the request.

unprocessedAccounts :: Maybe [UnprocessedAccount]

The requested member accounts for which Detective was unable to return member details.

For each account, provides the reason why the request could not be processed.

httpStatus :: Int

The response's http status code.

newGetMembersResponse

Arguments

:: Int
$sel:httpStatus:GetMembersResponse'

-> GetMembersResponse

Create a value of GetMembersResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:memberDetails:GetMembersResponse', getMembersResponse_memberDetails - The member account details that Detective is returning in response to the request.

$sel:unprocessedAccounts:GetMembersResponse', getMembersResponse_unprocessedAccounts - The requested member accounts for which Detective was unable to return member details.

For each account, provides the reason why the request could not be processed.

$sel:httpStatus:GetMembersResponse', getMembersResponse_httpStatus - The response's http status code.

Response Lenses

getMembersResponse_memberDetails :: Lens' GetMembersResponse (Maybe [MemberDetail])

The member account details that Detective is returning in response to the request.

getMembersResponse_unprocessedAccounts :: Lens' GetMembersResponse (Maybe [UnprocessedAccount])

The requested member accounts for which Detective was unable to return member details.

For each account, provides the reason why the request could not be processed.

getMembersResponse_httpStatus :: Lens' GetMembersResponse Int

The response's http status code.