Amazonka.Detective.Lens

Description

Operations

StartMonitoringMember

startMonitoringMember_graphArn :: Lens' StartMonitoringMember Text

The ARN of the behavior graph.

startMonitoringMember_accountId :: Lens' StartMonitoringMember Text

The account ID of the member account to try to enable.

The account must be an invited member account with a status of ACCEPTED_BUT_DISABLED.

DeleteMembers

deleteMembers_graphArn :: Lens' DeleteMembers Text

The ARN of the behavior graph to delete members from.

deleteMembers_accountIds :: Lens' DeleteMembers (NonEmpty Text)

The list of AWS account identifiers for the member accounts to delete from the behavior graph. You can delete up to 50 member accounts at a time.

deleteMembersResponse_accountIds :: Lens' DeleteMembersResponse (Maybe (NonEmpty Text))

The list of AWS account identifiers for the member accounts that Detective successfully deleted from the behavior graph.

deleteMembersResponse_unprocessedAccounts :: Lens' DeleteMembersResponse (Maybe [UnprocessedAccount])

The list of member accounts that Detective was not able to delete from the behavior graph. For each member account, provides the reason that the deletion could not be processed.

deleteMembersResponse_httpStatus :: Lens' DeleteMembersResponse Int

The response's http status code.

ListTagsForResource

listTagsForResource_resourceArn :: Lens' ListTagsForResource Text

The ARN of the behavior graph for which to retrieve the tag values.

listTagsForResourceResponse_tags :: Lens' ListTagsForResourceResponse (Maybe (HashMap Text Text))

The tag values that are assigned to the behavior graph. The request returns up to 50 tag values.

listTagsForResourceResponse_httpStatus :: Lens' ListTagsForResourceResponse Int

The response's http status code.

DeleteGraph

deleteGraph_graphArn :: Lens' DeleteGraph Text

The ARN of the behavior graph to disable.

ListInvitations

listInvitations_nextToken :: Lens' ListInvitations (Maybe Text)

For requests to retrieve the next page of results, the pagination token that was returned with the previous page of results. The initial request does not include a pagination token.

listInvitations_maxResults :: Lens' ListInvitations (Maybe Natural)

The maximum number of behavior graph invitations to return in the response. The total must be less than the overall limit on the number of results to return, which is currently 200.

listInvitationsResponse_invitations :: Lens' ListInvitationsResponse (Maybe [MemberDetail])

The list of behavior graphs for which the member account has open or accepted invitations.

listInvitationsResponse_nextToken :: Lens' ListInvitationsResponse (Maybe Text)

If there are more behavior graphs remaining in the results, then this is the pagination token to use to request the next page of behavior graphs.

listInvitationsResponse_httpStatus :: Lens' ListInvitationsResponse Int

The response's http status code.

DisassociateMembership

disassociateMembership_graphArn :: Lens' DisassociateMembership Text

The ARN of the behavior graph to remove the member account from.

The member account's member status in the behavior graph must be ENABLED.

AcceptInvitation

acceptInvitation_graphArn :: Lens' AcceptInvitation Text

The ARN of the behavior graph that the member account is accepting the invitation for.

The member account status in the behavior graph must be INVITED.

ListMembers

listMembers_nextToken :: Lens' ListMembers (Maybe Text)

For requests to retrieve the next page of member account results, the pagination token that was returned with the previous page of results. The initial request does not include a pagination token.

listMembers_maxResults :: Lens' ListMembers (Maybe Natural)

The maximum number of member accounts to include in the response. The total must be less than the overall limit on the number of results to return, which is currently 200.

listMembers_graphArn :: Lens' ListMembers Text

The ARN of the behavior graph for which to retrieve the list of member accounts.

listMembersResponse_memberDetails :: Lens' ListMembersResponse (Maybe [MemberDetail])

The list of member accounts in the behavior graph.

The results include member accounts that did not pass verification and member accounts that have not yet accepted the invitation to the behavior graph. The results do not include member accounts that were removed from the behavior graph.

listMembersResponse_nextToken :: Lens' ListMembersResponse (Maybe Text)

If there are more member accounts remaining in the results, then this is the pagination token to use to request the next page of member accounts.

listMembersResponse_httpStatus :: Lens' ListMembersResponse Int

The response's http status code.

CreateMembers

createMembers_disableEmailNotification :: Lens' CreateMembers (Maybe Bool)

if set to true, then the member accounts do not receive email notifications. By default, this is set to false, and the member accounts receive email notifications.

createMembers_message :: Lens' CreateMembers (Maybe Text)

Customized message text to include in the invitation email message to the invited member accounts.

createMembers_graphArn :: Lens' CreateMembers Text

The ARN of the behavior graph to invite the member accounts to contribute their data to.

createMembers_accounts :: Lens' CreateMembers (NonEmpty Account)

The list of AWS accounts to invite to become member accounts in the behavior graph. You can invite up to 50 accounts at a time. For each invited account, the account list contains the account identifier and the AWS account root user email address.

createMembersResponse_members :: Lens' CreateMembersResponse (Maybe [MemberDetail])

The set of member account invitation requests that Detective was able to process. This includes accounts that are being verified, that failed verification, and that passed verification and are being sent an invitation.

createMembersResponse_unprocessedAccounts :: Lens' CreateMembersResponse (Maybe [UnprocessedAccount])

The list of accounts for which Detective was unable to process the invitation request. For each account, the list provides the reason why the request could not be processed. The list includes accounts that are already member accounts in the behavior graph.

createMembersResponse_httpStatus :: Lens' CreateMembersResponse Int

The response's http status code.

GetMembers

getMembers_graphArn :: Lens' GetMembers Text

The ARN of the behavior graph for which to request the member details.

getMembers_accountIds :: Lens' GetMembers (NonEmpty Text)

The list of AWS account identifiers for the member account for which to return member details. You can request details for up to 50 member accounts at a time.

You cannot use GetMembers to retrieve information about member accounts that were removed from the behavior graph.

getMembersResponse_memberDetails :: Lens' GetMembersResponse (Maybe [MemberDetail])

The member account details that Detective is returning in response to the request.

getMembersResponse_unprocessedAccounts :: Lens' GetMembersResponse (Maybe [UnprocessedAccount])

The requested member accounts for which Detective was unable to return member details.

For each account, provides the reason why the request could not be processed.

getMembersResponse_httpStatus :: Lens' GetMembersResponse Int

The response's http status code.

ListGraphs

listGraphs_nextToken :: Lens' ListGraphs (Maybe Text)

For requests to get the next page of results, the pagination token that was returned with the previous set of results. The initial request does not include a pagination token.

listGraphs_maxResults :: Lens' ListGraphs (Maybe Natural)

The maximum number of graphs to return at a time. The total must be less than the overall limit on the number of results to return, which is currently 200.

listGraphsResponse_nextToken :: Lens' ListGraphsResponse (Maybe Text)

If there are more behavior graphs remaining in the results, then this is the pagination token to use to request the next page of behavior graphs.

listGraphsResponse_graphList :: Lens' ListGraphsResponse (Maybe [Graph])

A list of behavior graphs that the account is an administrator account for.

listGraphsResponse_httpStatus :: Lens' ListGraphsResponse Int

The response's http status code.

TagResource

tagResource_resourceArn :: Lens' TagResource Text

The ARN of the behavior graph to assign the tags to.

tagResource_tags :: Lens' TagResource (HashMap Text Text)

The tags to assign to the behavior graph. You can add up to 50 tags. For each tag, you provide the tag key and the tag value. Each tag key can contain up to 128 characters. Each tag value can contain up to 256 characters.

tagResourceResponse_httpStatus :: Lens' TagResourceResponse Int

The response's http status code.

CreateGraph

createGraph_tags :: Lens' CreateGraph (Maybe (HashMap Text Text))

The tags to assign to the new behavior graph. You can add up to 50 tags. For each tag, you provide the tag key and the tag value. Each tag key can contain up to 128 characters. Each tag value can contain up to 256 characters.

createGraphResponse_graphArn :: Lens' CreateGraphResponse (Maybe Text)

The ARN of the new behavior graph.

createGraphResponse_httpStatus :: Lens' CreateGraphResponse Int

The response's http status code.

UntagResource

untagResource_resourceArn :: Lens' UntagResource Text

The ARN of the behavior graph to remove the tags from.

untagResource_tagKeys :: Lens' UntagResource (NonEmpty Text)

The tag keys of the tags to remove from the behavior graph. You can remove up to 50 tags at a time.

untagResourceResponse_httpStatus :: Lens' UntagResourceResponse Int

The response's http status code.

RejectInvitation

rejectInvitation_graphArn :: Lens' RejectInvitation Text

The ARN of the behavior graph to reject the invitation to.

The member account's current member status in the behavior graph must be INVITED.

Types

Account

account_accountId :: Lens' Account Text

The account identifier of the AWS account.

account_emailAddress :: Lens' Account Text

The AWS account root user email address for the AWS account.

Graph

graph_arn :: Lens' Graph (Maybe Text)

The ARN of the behavior graph.

graph_createdTime :: Lens' Graph (Maybe UTCTime)

The date and time that the behavior graph was created. The value is in milliseconds since the epoch.

MemberDetail

memberDetail_percentOfGraphUtilizationUpdatedTime :: Lens' MemberDetail (Maybe UTCTime)

The date and time when the graph utilization percentage was last updated.

memberDetail_status :: Lens' MemberDetail (Maybe MemberStatus)

The current membership status of the member account. The status can have one of the following values:

INVITED - Indicates that the member was sent an invitation but has not yet responded.

VERIFICATION_IN_PROGRESS - Indicates that Detective is verifying that the account identifier and email address provided for the member account match. If they do match, then Detective sends the invitation. If the email address and account identifier don't match, then the member cannot be added to the behavior graph.

VERIFICATION_FAILED - Indicates that the account and email address provided for the member account do not match, and Detective did not send an invitation to the account.

ENABLED - Indicates that the member account accepted the invitation to contribute to the behavior graph.

ACCEPTED_BUT_DISABLED - Indicates that the member account accepted the invitation but is prevented from contributing data to the behavior graph. DisabledReason provides the reason why the member account is not enabled.

Member accounts that declined an invitation or that were removed from the behavior graph are not included.

memberDetail_invitedTime :: Lens' MemberDetail (Maybe UTCTime)

The date and time that Detective sent the invitation to the member account. The value is in milliseconds since the epoch.

memberDetail_administratorId :: Lens' MemberDetail (Maybe Text)

The AWS account identifier of the administrator account for the behavior graph.

memberDetail_graphArn :: Lens' MemberDetail (Maybe Text)

The ARN of the behavior graph that the member account was invited to.

memberDetail_masterId :: Lens' MemberDetail (Maybe Text)

The AWS account identifier of the administrator account for the behavior graph.

memberDetail_accountId :: Lens' MemberDetail (Maybe Text)

The AWS account identifier for the member account.

memberDetail_disabledReason :: Lens' MemberDetail (Maybe MemberDisabledReason)

For member accounts with a status of ACCEPTED_BUT_DISABLED, the reason that the member account is not enabled.

The reason can have one of the following values:

VOLUME_TOO_HIGH - Indicates that adding the member account would cause the data volume for the behavior graph to be too high.

VOLUME_UNKNOWN - Indicates that Detective is unable to verify the data volume for the member account. This is usually because the member account is not enrolled in Amazon GuardDuty.

memberDetail_percentOfGraphUtilization :: Lens' MemberDetail (Maybe Double)

The member account data volume as a percentage of the maximum allowed data volume. 0 indicates 0 percent, and 100 indicates 100 percent.

Note that this is not the percentage of the behavior graph data volume.

For example, the data volume for the behavior graph is 80 GB per day. The maximum data volume is 160 GB per day. If the data volume for the member account is 40 GB per day, then PercentOfGraphUtilization is 25. It represents 25% of the maximum allowed data volume.

memberDetail_emailAddress :: Lens' MemberDetail (Maybe Text)

The AWS account root user email address for the member account.

memberDetail_volumeUsageUpdatedTime :: Lens' MemberDetail (Maybe UTCTime)

The data and time when the member account data volume was last updated.

memberDetail_updatedTime :: Lens' MemberDetail (Maybe UTCTime)

The date and time that the member account was last updated. The value is in milliseconds since the epoch.

memberDetail_volumeUsageInBytes :: Lens' MemberDetail (Maybe Integer)

The data volume in bytes per day for the member account.

UnprocessedAccount

unprocessedAccount_accountId :: Lens' UnprocessedAccount (Maybe Text)

The AWS account identifier of the member account that was not processed.

unprocessedAccount_reason :: Lens' UnprocessedAccount (Maybe Text)

The reason that the member account request could not be processed.