Amazonka.Detective.ListGraphs

Description

Returns the list of behavior graphs that the calling account is an administrator account of. This operation can only be called by an administrator account.

Because an account can currently only be the administrator of one behavior graph within a Region, the results always contain a single behavior graph.

Creating a Request

data ListGraphs

See: newListGraphs smart constructor.

Constructors

ListGraphs'

Fields

nextToken :: Maybe Text

For requests to get the next page of results, the pagination token that was returned with the previous set of results. The initial request does not include a pagination token.

maxResults :: Maybe Natural

The maximum number of graphs to return at a time. The total must be less than the overall limit on the number of results to return, which is currently 200.

newListGraphs :: ListGraphs

Create a value of ListGraphs with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:nextToken:ListGraphs', listGraphs_nextToken - For requests to get the next page of results, the pagination token that was returned with the previous set of results. The initial request does not include a pagination token.

$sel:maxResults:ListGraphs', listGraphs_maxResults - The maximum number of graphs to return at a time. The total must be less than the overall limit on the number of results to return, which is currently 200.

Request Lenses

listGraphs_nextToken :: Lens' ListGraphs (Maybe Text)

For requests to get the next page of results, the pagination token that was returned with the previous set of results. The initial request does not include a pagination token.

listGraphs_maxResults :: Lens' ListGraphs (Maybe Natural)

The maximum number of graphs to return at a time. The total must be less than the overall limit on the number of results to return, which is currently 200.

Destructuring the Response

data ListGraphsResponse

See: newListGraphsResponse smart constructor.

Constructors

ListGraphsResponse'

Fields

nextToken :: Maybe Text

If there are more behavior graphs remaining in the results, then this is the pagination token to use to request the next page of behavior graphs.

graphList :: Maybe [Graph]

A list of behavior graphs that the account is an administrator account for.

httpStatus :: Int

The response's http status code.

newListGraphsResponse

Arguments

:: Int
$sel:httpStatus:ListGraphsResponse'

-> ListGraphsResponse

Create a value of ListGraphsResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:nextToken:ListGraphs', listGraphsResponse_nextToken - If there are more behavior graphs remaining in the results, then this is the pagination token to use to request the next page of behavior graphs.

$sel:graphList:ListGraphsResponse', listGraphsResponse_graphList - A list of behavior graphs that the account is an administrator account for.

$sel:httpStatus:ListGraphsResponse', listGraphsResponse_httpStatus - The response's http status code.

Response Lenses

listGraphsResponse_nextToken :: Lens' ListGraphsResponse (Maybe Text)

If there are more behavior graphs remaining in the results, then this is the pagination token to use to request the next page of behavior graphs.

listGraphsResponse_graphList :: Lens' ListGraphsResponse (Maybe [Graph])

A list of behavior graphs that the account is an administrator account for.

listGraphsResponse_httpStatus :: Lens' ListGraphsResponse Int

The response's http status code.