Amazonka.DevOpsGuru.AddNotificationChannel

Description

Adds a notification channel to DevOps Guru. A notification channel is used to notify you about important DevOps Guru events, such as when an insight is generated.

If you use an Amazon SNS topic in another account, you must attach a policy to it that grants DevOps Guru permission to it notifications. DevOps Guru adds the required policy on your behalf to send notifications using Amazon SNS in your account. For more information, see Permissions for cross account Amazon SNS topics.

If you use an Amazon SNS topic that is encrypted by an AWS Key Management Service customer-managed key (CMK), then you must add permissions to the CMK. For more information, see Permissions for AWS KMS–encrypted Amazon SNS topics.

Creating a Request

data AddNotificationChannel

See: newAddNotificationChannel smart constructor.

Constructors

AddNotificationChannel'

Fields

config :: NotificationChannelConfig

A NotificationChannelConfig object that specifies what type of notification channel to add. The one supported notification channel is Amazon Simple Notification Service (Amazon SNS).

newAddNotificationChannel

Arguments

:: NotificationChannelConfig
$sel:config:AddNotificationChannel'

-> AddNotificationChannel

Create a value of AddNotificationChannel with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:config:AddNotificationChannel', addNotificationChannel_config - A NotificationChannelConfig object that specifies what type of notification channel to add. The one supported notification channel is Amazon Simple Notification Service (Amazon SNS).

Request Lenses

addNotificationChannel_config :: Lens' AddNotificationChannel NotificationChannelConfig

A NotificationChannelConfig object that specifies what type of notification channel to add. The one supported notification channel is Amazon Simple Notification Service (Amazon SNS).

Destructuring the Response

data AddNotificationChannelResponse

See: newAddNotificationChannelResponse smart constructor.

Constructors

AddNotificationChannelResponse'

Fields

httpStatus :: Int

The response's http status code.

id :: Text

The ID of the added notification channel.

newAddNotificationChannelResponse

Arguments

:: Int
$sel:httpStatus:AddNotificationChannelResponse'

-> Text
$sel:id:AddNotificationChannelResponse'

-> AddNotificationChannelResponse

Create a value of AddNotificationChannelResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:httpStatus:AddNotificationChannelResponse', addNotificationChannelResponse_httpStatus - The response's http status code.

$sel:id:AddNotificationChannelResponse', addNotificationChannelResponse_id - The ID of the added notification channel.

Response Lenses

addNotificationChannelResponse_httpStatus :: Lens' AddNotificationChannelResponse Int

The response's http status code.

addNotificationChannelResponse_id :: Lens' AddNotificationChannelResponse Text

The ID of the added notification channel.