Amazonka.DevOpsGuru.DescribeAccountOverview

Description

For the time range passed in, returns the number of open reactive insight that were created, the number of open proactive insights that were created, and the Mean Time to Recover (MTTR) for all closed reactive insights.

Creating a Request

data DescribeAccountOverview

See: newDescribeAccountOverview smart constructor.

Constructors

DescribeAccountOverview'

Fields

toTime :: Maybe POSIX

The end of the time range passed in. The start time granularity is at the day level. The floor of the start time is used. Returned information occurred before this day. If this is not specified, then the current day is used.

fromTime :: POSIX

The start of the time range passed in. The start time granularity is at the day level. The floor of the start time is used. Returned information occurred after this day.

newDescribeAccountOverview

Arguments

:: UTCTime
$sel:fromTime:DescribeAccountOverview'

-> DescribeAccountOverview

Create a value of DescribeAccountOverview with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:toTime:DescribeAccountOverview', describeAccountOverview_toTime - The end of the time range passed in. The start time granularity is at the day level. The floor of the start time is used. Returned information occurred before this day. If this is not specified, then the current day is used.

$sel:fromTime:DescribeAccountOverview', describeAccountOverview_fromTime - The start of the time range passed in. The start time granularity is at the day level. The floor of the start time is used. Returned information occurred after this day.

Request Lenses

describeAccountOverview_toTime :: Lens' DescribeAccountOverview (Maybe UTCTime)

The end of the time range passed in. The start time granularity is at the day level. The floor of the start time is used. Returned information occurred before this day. If this is not specified, then the current day is used.

describeAccountOverview_fromTime :: Lens' DescribeAccountOverview UTCTime

The start of the time range passed in. The start time granularity is at the day level. The floor of the start time is used. Returned information occurred after this day.

Destructuring the Response

data DescribeAccountOverviewResponse

See: newDescribeAccountOverviewResponse smart constructor.

Constructors

DescribeAccountOverviewResponse'

Fields

httpStatus :: Int

The response's http status code.

reactiveInsights :: Int

An integer that specifies the number of open reactive insights in your AWS account that were created during the time range passed in.

proactiveInsights :: Int

An integer that specifies the number of open proactive insights in your AWS account that were created during the time range passed in.

meanTimeToRecoverInMilliseconds :: Integer

The Mean Time to Recover (MTTR) for all closed insights that were created during the time range passed in.

newDescribeAccountOverviewResponse

Arguments

:: Int
$sel:httpStatus:DescribeAccountOverviewResponse'

-> Int
$sel:reactiveInsights:DescribeAccountOverviewResponse'

-> Int
$sel:proactiveInsights:DescribeAccountOverviewResponse'

-> Integer
$sel:meanTimeToRecoverInMilliseconds:DescribeAccountOverviewResponse'

-> DescribeAccountOverviewResponse

Create a value of DescribeAccountOverviewResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:httpStatus:DescribeAccountOverviewResponse', describeAccountOverviewResponse_httpStatus - The response's http status code.

$sel:reactiveInsights:DescribeAccountOverviewResponse', describeAccountOverviewResponse_reactiveInsights - An integer that specifies the number of open reactive insights in your AWS account that were created during the time range passed in.

$sel:proactiveInsights:DescribeAccountOverviewResponse', describeAccountOverviewResponse_proactiveInsights - An integer that specifies the number of open proactive insights in your AWS account that were created during the time range passed in.

$sel:meanTimeToRecoverInMilliseconds:DescribeAccountOverviewResponse', describeAccountOverviewResponse_meanTimeToRecoverInMilliseconds - The Mean Time to Recover (MTTR) for all closed insights that were created during the time range passed in.

Response Lenses

describeAccountOverviewResponse_httpStatus :: Lens' DescribeAccountOverviewResponse Int

The response's http status code.

describeAccountOverviewResponse_reactiveInsights :: Lens' DescribeAccountOverviewResponse Int

An integer that specifies the number of open reactive insights in your AWS account that were created during the time range passed in.

describeAccountOverviewResponse_proactiveInsights :: Lens' DescribeAccountOverviewResponse Int

An integer that specifies the number of open proactive insights in your AWS account that were created during the time range passed in.

describeAccountOverviewResponse_meanTimeToRecoverInMilliseconds :: Lens' DescribeAccountOverviewResponse Integer

The Mean Time to Recover (MTTR) for all closed insights that were created during the time range passed in.