Amazonka.DevOpsGuru.DescribeResourceCollectionHealth

Description

Returns the number of open proactive insights, open reactive insights, and the Mean Time to Recover (MTTR) for all closed insights in resource collections in your account. You specify the type of AWS resources collection. The one type of AWS resource collection supported is AWS CloudFormation stacks. DevOps Guru can be configured to analyze only the AWS resources that are defined in the stacks. You can specify up to 500 AWS CloudFormation stacks.

This operation returns paginated results.

Creating a Request

data DescribeResourceCollectionHealth

See: newDescribeResourceCollectionHealth smart constructor.

Constructors

DescribeResourceCollectionHealth'

Fields

nextToken :: Maybe Text

The pagination token to use to retrieve the next page of results for this operation. If this value is null, it retrieves the first page.

resourceCollectionType :: ResourceCollectionType

An AWS resource collection type. This type specifies how analyzed AWS resources are defined. The one type of AWS resource collection supported is AWS CloudFormation stacks. DevOps Guru can be configured to analyze only the AWS resources that are defined in the stacks. You can specify up to 500 AWS CloudFormation stacks.

newDescribeResourceCollectionHealth

Arguments

:: ResourceCollectionType
$sel:resourceCollectionType:DescribeResourceCollectionHealth'

-> DescribeResourceCollectionHealth

Create a value of DescribeResourceCollectionHealth with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:nextToken:DescribeResourceCollectionHealth', describeResourceCollectionHealth_nextToken - The pagination token to use to retrieve the next page of results for this operation. If this value is null, it retrieves the first page.

$sel:resourceCollectionType:DescribeResourceCollectionHealth', describeResourceCollectionHealth_resourceCollectionType - An AWS resource collection type. This type specifies how analyzed AWS resources are defined. The one type of AWS resource collection supported is AWS CloudFormation stacks. DevOps Guru can be configured to analyze only the AWS resources that are defined in the stacks. You can specify up to 500 AWS CloudFormation stacks.

Request Lenses

describeResourceCollectionHealth_nextToken :: Lens' DescribeResourceCollectionHealth (Maybe Text)

The pagination token to use to retrieve the next page of results for this operation. If this value is null, it retrieves the first page.

describeResourceCollectionHealth_resourceCollectionType :: Lens' DescribeResourceCollectionHealth ResourceCollectionType

An AWS resource collection type. This type specifies how analyzed AWS resources are defined. The one type of AWS resource collection supported is AWS CloudFormation stacks. DevOps Guru can be configured to analyze only the AWS resources that are defined in the stacks. You can specify up to 500 AWS CloudFormation stacks.

Destructuring the Response

data DescribeResourceCollectionHealthResponse

See: newDescribeResourceCollectionHealthResponse smart constructor.

Constructors

DescribeResourceCollectionHealthResponse'

Fields

service :: Maybe [ServiceHealth]

An array of ServiceHealth objects that describes the health of the AWS services associated with the resources in the collection.

nextToken :: Maybe Text

The pagination token to use to retrieve the next page of results for this operation. If there are no more pages, this value is null.

httpStatus :: Int

The response's http status code.

cloudFormation :: [CloudFormationHealth]

The returned CloudFormationHealthOverview object that contains an InsightHealthOverview object with the requested system health information.

newDescribeResourceCollectionHealthResponse

Arguments

:: Int
$sel:httpStatus:DescribeResourceCollectionHealthResponse'

-> DescribeResourceCollectionHealthResponse

Create a value of DescribeResourceCollectionHealthResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:service:DescribeResourceCollectionHealthResponse', describeResourceCollectionHealthResponse_service - An array of ServiceHealth objects that describes the health of the AWS services associated with the resources in the collection.

$sel:nextToken:DescribeResourceCollectionHealth', describeResourceCollectionHealthResponse_nextToken - The pagination token to use to retrieve the next page of results for this operation. If there are no more pages, this value is null.

$sel:httpStatus:DescribeResourceCollectionHealthResponse', describeResourceCollectionHealthResponse_httpStatus - The response's http status code.

$sel:cloudFormation:DescribeResourceCollectionHealthResponse', describeResourceCollectionHealthResponse_cloudFormation - The returned CloudFormationHealthOverview object that contains an InsightHealthOverview object with the requested system health information.

Response Lenses

describeResourceCollectionHealthResponse_service :: Lens' DescribeResourceCollectionHealthResponse (Maybe [ServiceHealth])

An array of ServiceHealth objects that describes the health of the AWS services associated with the resources in the collection.

describeResourceCollectionHealthResponse_nextToken :: Lens' DescribeResourceCollectionHealthResponse (Maybe Text)

The pagination token to use to retrieve the next page of results for this operation. If there are no more pages, this value is null.

describeResourceCollectionHealthResponse_httpStatus :: Lens' DescribeResourceCollectionHealthResponse Int

The response's http status code.

describeResourceCollectionHealthResponse_cloudFormation :: Lens' DescribeResourceCollectionHealthResponse [CloudFormationHealth]

The returned CloudFormationHealthOverview object that contains an InsightHealthOverview object with the requested system health information.