Amazonka.DevOpsGuru.DescribeServiceIntegration

Description

Returns the integration status of services that are integrated with DevOps Guru. The one service that can be integrated with DevOps Guru is AWS Systems Manager, which can be used to create an OpsItem for each generated insight.

Creating a Request

data DescribeServiceIntegration

See: newDescribeServiceIntegration smart constructor.

Constructors

DescribeServiceIntegration'

newDescribeServiceIntegration :: DescribeServiceIntegration

Create a value of DescribeServiceIntegration with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

Destructuring the Response

data DescribeServiceIntegrationResponse

See: newDescribeServiceIntegrationResponse smart constructor.

Constructors

DescribeServiceIntegrationResponse'

Fields

serviceIntegration :: Maybe ServiceIntegrationConfig

httpStatus :: Int

The response's http status code.

newDescribeServiceIntegrationResponse

Arguments

:: Int
$sel:httpStatus:DescribeServiceIntegrationResponse'

-> DescribeServiceIntegrationResponse

Create a value of DescribeServiceIntegrationResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:serviceIntegration:DescribeServiceIntegrationResponse', describeServiceIntegrationResponse_serviceIntegration - Undocumented member.

$sel:httpStatus:DescribeServiceIntegrationResponse', describeServiceIntegrationResponse_httpStatus - The response's http status code.

Response Lenses

describeServiceIntegrationResponse_serviceIntegration :: Lens' DescribeServiceIntegrationResponse (Maybe ServiceIntegrationConfig)

Undocumented member.

describeServiceIntegrationResponse_httpStatus :: Lens' DescribeServiceIntegrationResponse Int

The response's http status code.