Amazonka.DevOpsGuru.GetResourceCollection

Description

Returns lists AWS resources that are of the specified resource collection type. The one type of AWS resource collection supported is AWS CloudFormation stacks. DevOps Guru can be configured to analyze only the AWS resources that are defined in the stacks. You can specify up to 500 AWS CloudFormation stacks.

This operation returns paginated results.

Creating a Request

data GetResourceCollection

See: newGetResourceCollection smart constructor.

Constructors

GetResourceCollection'

Fields

nextToken :: Maybe Text

The pagination token to use to retrieve the next page of results for this operation. If this value is null, it retrieves the first page.

resourceCollectionType :: ResourceCollectionType

The type of AWS resource collections to return. The one valid value is CLOUD_FORMATION for AWS CloudFormation stacks.

newGetResourceCollection

Arguments

:: ResourceCollectionType
$sel:resourceCollectionType:GetResourceCollection'

-> GetResourceCollection

Create a value of GetResourceCollection with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:nextToken:GetResourceCollection', getResourceCollection_nextToken - The pagination token to use to retrieve the next page of results for this operation. If this value is null, it retrieves the first page.

$sel:resourceCollectionType:GetResourceCollection', getResourceCollection_resourceCollectionType - The type of AWS resource collections to return. The one valid value is CLOUD_FORMATION for AWS CloudFormation stacks.

Request Lenses

getResourceCollection_nextToken :: Lens' GetResourceCollection (Maybe Text)

The pagination token to use to retrieve the next page of results for this operation. If this value is null, it retrieves the first page.

getResourceCollection_resourceCollectionType :: Lens' GetResourceCollection ResourceCollectionType

The type of AWS resource collections to return. The one valid value is CLOUD_FORMATION for AWS CloudFormation stacks.

Destructuring the Response

data GetResourceCollectionResponse

See: newGetResourceCollectionResponse smart constructor.

Constructors

GetResourceCollectionResponse'

Fields

resourceCollection :: Maybe ResourceCollectionFilter

The requested list of AWS resource collections. The one type of AWS resource collection supported is AWS CloudFormation stacks. DevOps Guru can be configured to analyze only the AWS resources that are defined in the stacks. You can specify up to 500 AWS CloudFormation stacks.

nextToken :: Maybe Text

The pagination token to use to retrieve the next page of results for this operation. If there are no more pages, this value is null.

httpStatus :: Int

The response's http status code.

newGetResourceCollectionResponse

Arguments

:: Int
$sel:httpStatus:GetResourceCollectionResponse'

-> GetResourceCollectionResponse

Create a value of GetResourceCollectionResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:resourceCollection:GetResourceCollectionResponse', getResourceCollectionResponse_resourceCollection - The requested list of AWS resource collections. The one type of AWS resource collection supported is AWS CloudFormation stacks. DevOps Guru can be configured to analyze only the AWS resources that are defined in the stacks. You can specify up to 500 AWS CloudFormation stacks.

$sel:nextToken:GetResourceCollection', getResourceCollectionResponse_nextToken - The pagination token to use to retrieve the next page of results for this operation. If there are no more pages, this value is null.

$sel:httpStatus:GetResourceCollectionResponse', getResourceCollectionResponse_httpStatus - The response's http status code.

Response Lenses

getResourceCollectionResponse_resourceCollection :: Lens' GetResourceCollectionResponse (Maybe ResourceCollectionFilter)

The requested list of AWS resource collections. The one type of AWS resource collection supported is AWS CloudFormation stacks. DevOps Guru can be configured to analyze only the AWS resources that are defined in the stacks. You can specify up to 500 AWS CloudFormation stacks.

getResourceCollectionResponse_nextToken :: Lens' GetResourceCollectionResponse (Maybe Text)

The pagination token to use to retrieve the next page of results for this operation. If there are no more pages, this value is null.

getResourceCollectionResponse_httpStatus :: Lens' GetResourceCollectionResponse Int

The response's http status code.