Amazonka.DevOpsGuru.ListRecommendations

Description

Returns a list of a specified insight's recommendations. Each recommendation includes a list of related metrics and a list of related events.

This operation returns paginated results.

Creating a Request

data ListRecommendations

See: newListRecommendations smart constructor.

Constructors

ListRecommendations'

Fields

locale :: Maybe Locale

A locale that specifies the language to use for recommendations.

nextToken :: Maybe Text

The pagination token to use to retrieve the next page of results for this operation. If this value is null, it retrieves the first page.

insightId :: Text

The ID of the requested insight.

newListRecommendations

Arguments

:: Text
$sel:insightId:ListRecommendations'

-> ListRecommendations

Create a value of ListRecommendations with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:locale:ListRecommendations', listRecommendations_locale - A locale that specifies the language to use for recommendations.

$sel:nextToken:ListRecommendations', listRecommendations_nextToken - The pagination token to use to retrieve the next page of results for this operation. If this value is null, it retrieves the first page.

$sel:insightId:ListRecommendations', listRecommendations_insightId - The ID of the requested insight.

Request Lenses

listRecommendations_locale :: Lens' ListRecommendations (Maybe Locale)

A locale that specifies the language to use for recommendations.

listRecommendations_nextToken :: Lens' ListRecommendations (Maybe Text)

The pagination token to use to retrieve the next page of results for this operation. If this value is null, it retrieves the first page.

listRecommendations_insightId :: Lens' ListRecommendations Text

The ID of the requested insight.

Destructuring the Response

data ListRecommendationsResponse

See: newListRecommendationsResponse smart constructor.

Constructors

ListRecommendationsResponse'

Fields

nextToken :: Maybe Text

The pagination token to use to retrieve the next page of results for this operation. If there are no more pages, this value is null.

recommendations :: Maybe [Recommendation]

An array of the requested recommendations.

httpStatus :: Int

The response's http status code.

newListRecommendationsResponse

Arguments

:: Int
$sel:httpStatus:ListRecommendationsResponse'

-> ListRecommendationsResponse

Create a value of ListRecommendationsResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:nextToken:ListRecommendations', listRecommendationsResponse_nextToken - The pagination token to use to retrieve the next page of results for this operation. If there are no more pages, this value is null.

$sel:recommendations:ListRecommendationsResponse', listRecommendationsResponse_recommendations - An array of the requested recommendations.

$sel:httpStatus:ListRecommendationsResponse', listRecommendationsResponse_httpStatus - The response's http status code.

Response Lenses

listRecommendationsResponse_nextToken :: Lens' ListRecommendationsResponse (Maybe Text)

The pagination token to use to retrieve the next page of results for this operation. If there are no more pages, this value is null.

listRecommendationsResponse_recommendations :: Lens' ListRecommendationsResponse (Maybe [Recommendation])

An array of the requested recommendations.

listRecommendationsResponse_httpStatus :: Lens' ListRecommendationsResponse Int

The response's http status code.