Amazonka.DevOpsGuru.SearchInsights

Description

Returns a list of insights in your AWS account. You can specify which insights are returned by their start time, one or more statuses (ONGOING, CLOSED, and CLOSED), one or more severities (LOW, MEDIUM, and HIGH), and type (REACTIVE or PROACTIVE).

Use the Filters parameter to specify status and severity search parameters. Use the Type parameter to specify REACTIVE or PROACTIVE in your search.

This operation returns paginated results.

Creating a Request

data SearchInsights

See: newSearchInsights smart constructor.

Constructors

SearchInsights'

Fields

filters :: Maybe SearchInsightsFilters

A SearchInsightsFilters object that is used to set the severity and status filters on your insight search.

nextToken :: Maybe Text

The pagination token to use to retrieve the next page of results for this operation. If this value is null, it retrieves the first page.

maxResults :: Maybe Natural

The maximum number of results to return with a single call. To retrieve the remaining results, make another call with the returned nextToken value.

startTimeRange :: StartTimeRange

The start of the time range passed in. Returned insights occurred after this time.

type' :: InsightType

The type of insights you are searching for (REACTIVE or PROACTIVE).

newSearchInsights

Arguments

:: StartTimeRange
$sel:startTimeRange:SearchInsights'

-> InsightType
$sel:type':SearchInsights'

-> SearchInsights

Create a value of SearchInsights with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:filters:SearchInsights', searchInsights_filters - A SearchInsightsFilters object that is used to set the severity and status filters on your insight search.

$sel:nextToken:SearchInsights', searchInsights_nextToken - The pagination token to use to retrieve the next page of results for this operation. If this value is null, it retrieves the first page.

$sel:maxResults:SearchInsights', searchInsights_maxResults - The maximum number of results to return with a single call. To retrieve the remaining results, make another call with the returned nextToken value.

$sel:startTimeRange:SearchInsights', searchInsights_startTimeRange - The start of the time range passed in. Returned insights occurred after this time.

$sel:type':SearchInsights', searchInsights_type - The type of insights you are searching for (REACTIVE or PROACTIVE).

Request Lenses

searchInsights_filters :: Lens' SearchInsights (Maybe SearchInsightsFilters)

A SearchInsightsFilters object that is used to set the severity and status filters on your insight search.

searchInsights_nextToken :: Lens' SearchInsights (Maybe Text)

The pagination token to use to retrieve the next page of results for this operation. If this value is null, it retrieves the first page.

searchInsights_maxResults :: Lens' SearchInsights (Maybe Natural)

The maximum number of results to return with a single call. To retrieve the remaining results, make another call with the returned nextToken value.

searchInsights_startTimeRange :: Lens' SearchInsights StartTimeRange

The start of the time range passed in. Returned insights occurred after this time.

searchInsights_type :: Lens' SearchInsights InsightType

The type of insights you are searching for (REACTIVE or PROACTIVE).

Destructuring the Response

data SearchInsightsResponse

See: newSearchInsightsResponse smart constructor.

Constructors

SearchInsightsResponse'

Fields

reactiveInsights :: Maybe [ReactiveInsightSummary]

The returned reactive insights.

nextToken :: Maybe Text

The pagination token to use to retrieve the next page of results for this operation. If there are no more pages, this value is null.

proactiveInsights :: Maybe [ProactiveInsightSummary]

The returned proactive insights.

httpStatus :: Int

The response's http status code.

newSearchInsightsResponse

Arguments

:: Int
$sel:httpStatus:SearchInsightsResponse'

-> SearchInsightsResponse

Create a value of SearchInsightsResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:reactiveInsights:SearchInsightsResponse', searchInsightsResponse_reactiveInsights - The returned reactive insights.

$sel:nextToken:SearchInsights', searchInsightsResponse_nextToken - The pagination token to use to retrieve the next page of results for this operation. If there are no more pages, this value is null.

$sel:proactiveInsights:SearchInsightsResponse', searchInsightsResponse_proactiveInsights - The returned proactive insights.

$sel:httpStatus:SearchInsightsResponse', searchInsightsResponse_httpStatus - The response's http status code.

Response Lenses

searchInsightsResponse_reactiveInsights :: Lens' SearchInsightsResponse (Maybe [ReactiveInsightSummary])

The returned reactive insights.

searchInsightsResponse_nextToken :: Lens' SearchInsightsResponse (Maybe Text)

The pagination token to use to retrieve the next page of results for this operation. If there are no more pages, this value is null.

searchInsightsResponse_proactiveInsights :: Lens' SearchInsightsResponse (Maybe [ProactiveInsightSummary])

The returned proactive insights.

searchInsightsResponse_httpStatus :: Lens' SearchInsightsResponse Int

The response's http status code.