Amazonka.DevOpsGuru.Types.CloudWatchMetricsDetail

Description

Documentation

data CloudWatchMetricsDetail

Information about an Amazon CloudWatch metric.

See: newCloudWatchMetricsDetail smart constructor.

Constructors

CloudWatchMetricsDetail'

Fields

period :: Maybe Int

The length of time associated with the CloudWatch metric in number of seconds.

metricName :: Maybe Text

The name of the CloudWatch metric.

namespace :: Maybe Text

The namespace of the CloudWatch metric. A namespace is a container for CloudWatch metrics.

stat :: Maybe CloudWatchMetricsStat

The type of statistic associated with the CloudWatch metric. For more information, see Statistics in the Amazon CloudWatch User Guide.

dimensions :: Maybe [CloudWatchMetricsDimension]

An array of CloudWatch dimensions associated with

unit :: Maybe Text

The unit of measure used for the CloudWatch metric. For example, Bytes, Seconds, Count, and Percent.

newCloudWatchMetricsDetail :: CloudWatchMetricsDetail

Create a value of CloudWatchMetricsDetail with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:period:CloudWatchMetricsDetail', cloudWatchMetricsDetail_period - The length of time associated with the CloudWatch metric in number of seconds.

$sel:metricName:CloudWatchMetricsDetail', cloudWatchMetricsDetail_metricName - The name of the CloudWatch metric.

$sel:namespace:CloudWatchMetricsDetail', cloudWatchMetricsDetail_namespace - The namespace of the CloudWatch metric. A namespace is a container for CloudWatch metrics.

$sel:stat:CloudWatchMetricsDetail', cloudWatchMetricsDetail_stat - The type of statistic associated with the CloudWatch metric. For more information, see Statistics in the Amazon CloudWatch User Guide.

$sel:dimensions:CloudWatchMetricsDetail', cloudWatchMetricsDetail_dimensions - An array of CloudWatch dimensions associated with

$sel:unit:CloudWatchMetricsDetail', cloudWatchMetricsDetail_unit - The unit of measure used for the CloudWatch metric. For example, Bytes, Seconds, Count, and Percent.

cloudWatchMetricsDetail_period :: Lens' CloudWatchMetricsDetail (Maybe Int)

The length of time associated with the CloudWatch metric in number of seconds.

cloudWatchMetricsDetail_metricName :: Lens' CloudWatchMetricsDetail (Maybe Text)

The name of the CloudWatch metric.

cloudWatchMetricsDetail_namespace :: Lens' CloudWatchMetricsDetail (Maybe Text)

The namespace of the CloudWatch metric. A namespace is a container for CloudWatch metrics.

cloudWatchMetricsDetail_stat :: Lens' CloudWatchMetricsDetail (Maybe CloudWatchMetricsStat)

The type of statistic associated with the CloudWatch metric. For more information, see Statistics in the Amazon CloudWatch User Guide.

cloudWatchMetricsDetail_dimensions :: Lens' CloudWatchMetricsDetail (Maybe [CloudWatchMetricsDimension])

An array of CloudWatch dimensions associated with

cloudWatchMetricsDetail_unit :: Lens' CloudWatchMetricsDetail (Maybe Text)

The unit of measure used for the CloudWatch metric. For example, Bytes, Seconds, Count, and Percent.