Amazonka.DevOpsGuru.Types.ListEventsFilters

Description

Documentation

data ListEventsFilters

Filters you can use to specify which events are returned when ListEvents is called.

See: newListEventsFilters smart constructor.

Constructors

ListEventsFilters'

Fields

resourceCollection :: Maybe ResourceCollection

eventClass :: Maybe EventClass

The class of the events you want to filter for, such as an infrastructure change, a deployment, or a schema change.

insightId :: Maybe Text

An ID of an insight that is related to the events you want to filter for.

dataSource :: Maybe EventDataSource

The source, AWS_CLOUD_TRAIL or AWS_CODE_DEPLOY, of the events you want returned.

eventTimeRange :: Maybe EventTimeRange

A time range during which you want the filtered events to have occurred.

eventSource :: Maybe Text

The AWS source that emitted the events you want to filter for.

newListEventsFilters :: ListEventsFilters

Create a value of ListEventsFilters with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:resourceCollection:ListEventsFilters', listEventsFilters_resourceCollection - Undocumented member.

$sel:eventClass:ListEventsFilters', listEventsFilters_eventClass - The class of the events you want to filter for, such as an infrastructure change, a deployment, or a schema change.

$sel:insightId:ListEventsFilters', listEventsFilters_insightId - An ID of an insight that is related to the events you want to filter for.

$sel:dataSource:ListEventsFilters', listEventsFilters_dataSource - The source, AWS_CLOUD_TRAIL or AWS_CODE_DEPLOY, of the events you want returned.

$sel:eventTimeRange:ListEventsFilters', listEventsFilters_eventTimeRange - A time range during which you want the filtered events to have occurred.

$sel:eventSource:ListEventsFilters', listEventsFilters_eventSource - The AWS source that emitted the events you want to filter for.

listEventsFilters_resourceCollection :: Lens' ListEventsFilters (Maybe ResourceCollection)

Undocumented member.

listEventsFilters_eventClass :: Lens' ListEventsFilters (Maybe EventClass)

The class of the events you want to filter for, such as an infrastructure change, a deployment, or a schema change.

listEventsFilters_insightId :: Lens' ListEventsFilters (Maybe Text)

An ID of an insight that is related to the events you want to filter for.

listEventsFilters_dataSource :: Lens' ListEventsFilters (Maybe EventDataSource)

The source, AWS_CLOUD_TRAIL or AWS_CODE_DEPLOY, of the events you want returned.

listEventsFilters_eventTimeRange :: Lens' ListEventsFilters (Maybe EventTimeRange)

A time range during which you want the filtered events to have occurred.

listEventsFilters_eventSource :: Lens' ListEventsFilters (Maybe Text)

The AWS source that emitted the events you want to filter for.