Amazonka.DevOpsGuru.Types.ListInsightsStatusFilter

Description

Documentation

data ListInsightsStatusFilter

A filter used by ListInsights to specify which insights to return.

See: newListInsightsStatusFilter smart constructor.

Constructors

ListInsightsStatusFilter'

Fields

closed :: Maybe ListInsightsClosedStatusFilter

A ListInsightsClosedStatusFilter that specifies closed insights that are either REACTIVE or PROACTIVE.

ongoing :: Maybe ListInsightsOngoingStatusFilter

A ListInsightsAnyStatusFilter that specifies ongoing insights that are either REACTIVE or PROACTIVE.

any :: Maybe ListInsightsAnyStatusFilter

A ListInsightsAnyStatusFilter that specifies insights of any status that are either REACTIVE or PROACTIVE.

newListInsightsStatusFilter :: ListInsightsStatusFilter

Create a value of ListInsightsStatusFilter with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:closed:ListInsightsStatusFilter', listInsightsStatusFilter_closed - A ListInsightsClosedStatusFilter that specifies closed insights that are either REACTIVE or PROACTIVE.

$sel:ongoing:ListInsightsStatusFilter', listInsightsStatusFilter_ongoing - A ListInsightsAnyStatusFilter that specifies ongoing insights that are either REACTIVE or PROACTIVE.

$sel:any:ListInsightsStatusFilter', listInsightsStatusFilter_any - A ListInsightsAnyStatusFilter that specifies insights of any status that are either REACTIVE or PROACTIVE.

listInsightsStatusFilter_closed :: Lens' ListInsightsStatusFilter (Maybe ListInsightsClosedStatusFilter)

A ListInsightsClosedStatusFilter that specifies closed insights that are either REACTIVE or PROACTIVE.

listInsightsStatusFilter_ongoing :: Lens' ListInsightsStatusFilter (Maybe ListInsightsOngoingStatusFilter)

A ListInsightsAnyStatusFilter that specifies ongoing insights that are either REACTIVE or PROACTIVE.

listInsightsStatusFilter_any :: Lens' ListInsightsStatusFilter (Maybe ListInsightsAnyStatusFilter)

A ListInsightsAnyStatusFilter that specifies insights of any status that are either REACTIVE or PROACTIVE.