Amazonka.DevOpsGuru.Types.ProactiveAnomalySummary

Description

Documentation

data ProactiveAnomalySummary

Details about a proactive anomaly. This object is returned by DescribeAnomaly.

See: newProactiveAnomalySummary smart constructor.

Constructors

ProactiveAnomalySummary'

Fields

anomalyReportedTimeRange :: Maybe AnomalyReportedTimeRange

A AnomalyReportedTimeRange object that specifies the time range between when the anomaly is opened and the time when it is closed.

status :: Maybe AnomalyStatus

The status of the anomaly.

resourceCollection :: Maybe ResourceCollection

severity :: Maybe AnomalySeverity

The severity of the anomaly.

updateTime :: Maybe POSIX

The time of the anomaly's most recent update.

sourceDetails :: Maybe AnomalySourceDetails

Details about the source of the analyzed operational data that triggered the anomaly. The one supported source is Amazon CloudWatch metrics.

predictionTimeRange :: Maybe PredictionTimeRange

limit :: Maybe Double

A threshold that was exceeded by behavior in analyzed resources. Exceeding this threshold is related to the anomalous behavior that generated this anomaly.

id :: Maybe Text

The ID of the anomaly.

associatedInsightId :: Maybe Text

The ID of the insight that contains this anomaly. An insight is composed of related anomalies.

anomalyTimeRange :: Maybe AnomalyTimeRange

newProactiveAnomalySummary :: ProactiveAnomalySummary

Create a value of ProactiveAnomalySummary with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:anomalyReportedTimeRange:ProactiveAnomalySummary', proactiveAnomalySummary_anomalyReportedTimeRange - A AnomalyReportedTimeRange object that specifies the time range between when the anomaly is opened and the time when it is closed.

$sel:status:ProactiveAnomalySummary', proactiveAnomalySummary_status - The status of the anomaly.

$sel:resourceCollection:ProactiveAnomalySummary', proactiveAnomalySummary_resourceCollection - Undocumented member.

$sel:severity:ProactiveAnomalySummary', proactiveAnomalySummary_severity - The severity of the anomaly.

$sel:updateTime:ProactiveAnomalySummary', proactiveAnomalySummary_updateTime - The time of the anomaly's most recent update.

$sel:sourceDetails:ProactiveAnomalySummary', proactiveAnomalySummary_sourceDetails - Details about the source of the analyzed operational data that triggered the anomaly. The one supported source is Amazon CloudWatch metrics.

$sel:predictionTimeRange:ProactiveAnomalySummary', proactiveAnomalySummary_predictionTimeRange - Undocumented member.

$sel:limit:ProactiveAnomalySummary', proactiveAnomalySummary_limit - A threshold that was exceeded by behavior in analyzed resources. Exceeding this threshold is related to the anomalous behavior that generated this anomaly.

$sel:id:ProactiveAnomalySummary', proactiveAnomalySummary_id - The ID of the anomaly.

$sel:associatedInsightId:ProactiveAnomalySummary', proactiveAnomalySummary_associatedInsightId - The ID of the insight that contains this anomaly. An insight is composed of related anomalies.

$sel:anomalyTimeRange:ProactiveAnomalySummary', proactiveAnomalySummary_anomalyTimeRange - Undocumented member.

proactiveAnomalySummary_anomalyReportedTimeRange :: Lens' ProactiveAnomalySummary (Maybe AnomalyReportedTimeRange)

A AnomalyReportedTimeRange object that specifies the time range between when the anomaly is opened and the time when it is closed.

proactiveAnomalySummary_status :: Lens' ProactiveAnomalySummary (Maybe AnomalyStatus)

The status of the anomaly.

proactiveAnomalySummary_resourceCollection :: Lens' ProactiveAnomalySummary (Maybe ResourceCollection)

Undocumented member.

proactiveAnomalySummary_severity :: Lens' ProactiveAnomalySummary (Maybe AnomalySeverity)

The severity of the anomaly.

proactiveAnomalySummary_updateTime :: Lens' ProactiveAnomalySummary (Maybe UTCTime)

The time of the anomaly's most recent update.

proactiveAnomalySummary_sourceDetails :: Lens' ProactiveAnomalySummary (Maybe AnomalySourceDetails)

Details about the source of the analyzed operational data that triggered the anomaly. The one supported source is Amazon CloudWatch metrics.

proactiveAnomalySummary_predictionTimeRange :: Lens' ProactiveAnomalySummary (Maybe PredictionTimeRange)

Undocumented member.

proactiveAnomalySummary_limit :: Lens' ProactiveAnomalySummary (Maybe Double)

A threshold that was exceeded by behavior in analyzed resources. Exceeding this threshold is related to the anomalous behavior that generated this anomaly.

proactiveAnomalySummary_id :: Lens' ProactiveAnomalySummary (Maybe Text)

The ID of the anomaly.

proactiveAnomalySummary_associatedInsightId :: Lens' ProactiveAnomalySummary (Maybe Text)

The ID of the insight that contains this anomaly. An insight is composed of related anomalies.

proactiveAnomalySummary_anomalyTimeRange :: Lens' ProactiveAnomalySummary (Maybe AnomalyTimeRange)

Undocumented member.