Amazonka.DevOpsGuru.Types.ResourceCollectionFilter

Description

Documentation

data ResourceCollectionFilter

Information about a filter used to specify which AWS resources are analyzed for anomalous behavior by DevOps Guru.

See: newResourceCollectionFilter smart constructor.

Constructors

ResourceCollectionFilter'

Fields

cloudFormation :: Maybe CloudFormationCollectionFilter

Information about AWS CloudFormation stacks. You can use up to 500 stacks to specify which AWS resources in your account to analyze. For more information, see Stacks in the AWS CloudFormation User Guide.

newResourceCollectionFilter :: ResourceCollectionFilter

Create a value of ResourceCollectionFilter with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:cloudFormation:ResourceCollectionFilter', resourceCollectionFilter_cloudFormation - Information about AWS CloudFormation stacks. You can use up to 500 stacks to specify which AWS resources in your account to analyze. For more information, see Stacks in the AWS CloudFormation User Guide.

resourceCollectionFilter_cloudFormation :: Lens' ResourceCollectionFilter (Maybe CloudFormationCollectionFilter)

Information about AWS CloudFormation stacks. You can use up to 500 stacks to specify which AWS resources in your account to analyze. For more information, see Stacks in the AWS CloudFormation User Guide.