Amazonka.DevOpsGuru.Types.StartTimeRange

Description

Documentation

data StartTimeRange

A time range used to specify when the behavior of an insight or anomaly started.

See: newStartTimeRange smart constructor.

Constructors

StartTimeRange'

Fields

fromTime :: Maybe POSIX

The start time of the time range.

toTime :: Maybe POSIX

The end time of the time range.

newStartTimeRange :: StartTimeRange

Create a value of StartTimeRange with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:fromTime:StartTimeRange', startTimeRange_fromTime - The start time of the time range.

$sel:toTime:StartTimeRange', startTimeRange_toTime - The end time of the time range.

startTimeRange_fromTime :: Lens' StartTimeRange (Maybe UTCTime)

The start time of the time range.

startTimeRange_toTime :: Lens' StartTimeRange (Maybe UTCTime)

The end time of the time range.