Amazonka.DeviceFarm.GetDevicePoolCompatibility

Description

Gets information about compatibility with a device pool.

Creating a Request

data GetDevicePoolCompatibility

Represents a request to the get device pool compatibility operation.

See: newGetDevicePoolCompatibility smart constructor.

Constructors

GetDevicePoolCompatibility'

Fields

test :: Maybe ScheduleRunTest

Information about the uploaded test to be run against the device pool.

appArn :: Maybe Text

The ARN of the app that is associated with the specified device pool.

configuration :: Maybe ScheduleRunConfiguration

An object that contains information about the settings for a run.

testType :: Maybe TestType

The test type for the specified device pool.

Allowed values include the following:

BUILTIN_FUZZ.

BUILTIN_EXPLORER. For Android, an app explorer that traverses an Android app, interacting with it and capturing screenshots at the same time.

APPIUM_JAVA_JUNIT.

APPIUM_JAVA_TESTNG.

APPIUM_PYTHON.

APPIUM_NODE.

APPIUM_RUBY.

APPIUM_WEB_JAVA_JUNIT.

APPIUM_WEB_JAVA_TESTNG.

APPIUM_WEB_PYTHON.

APPIUM_WEB_NODE.

APPIUM_WEB_RUBY.

CALABASH.

INSTRUMENTATION.

UIAUTOMATION.

UIAUTOMATOR.

XCTEST.

XCTEST_UI.

devicePoolArn :: Text

The device pool's ARN.

newGetDevicePoolCompatibility

Arguments

:: Text
$sel:devicePoolArn:GetDevicePoolCompatibility'

-> GetDevicePoolCompatibility

Create a value of GetDevicePoolCompatibility with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:test:GetDevicePoolCompatibility', getDevicePoolCompatibility_test - Information about the uploaded test to be run against the device pool.

$sel:appArn:GetDevicePoolCompatibility', getDevicePoolCompatibility_appArn - The ARN of the app that is associated with the specified device pool.

$sel:configuration:GetDevicePoolCompatibility', getDevicePoolCompatibility_configuration - An object that contains information about the settings for a run.

$sel:testType:GetDevicePoolCompatibility', getDevicePoolCompatibility_testType - The test type for the specified device pool.

Allowed values include the following:

BUILTIN_FUZZ.

BUILTIN_EXPLORER. For Android, an app explorer that traverses an Android app, interacting with it and capturing screenshots at the same time.

APPIUM_JAVA_JUNIT.

APPIUM_JAVA_TESTNG.

APPIUM_PYTHON.

APPIUM_NODE.

APPIUM_RUBY.

APPIUM_WEB_JAVA_JUNIT.

APPIUM_WEB_JAVA_TESTNG.

APPIUM_WEB_PYTHON.

APPIUM_WEB_NODE.

APPIUM_WEB_RUBY.

CALABASH.

INSTRUMENTATION.

UIAUTOMATION.

UIAUTOMATOR.

XCTEST.

XCTEST_UI.

$sel:devicePoolArn:GetDevicePoolCompatibility', getDevicePoolCompatibility_devicePoolArn - The device pool's ARN.

Request Lenses

getDevicePoolCompatibility_test :: Lens' GetDevicePoolCompatibility (Maybe ScheduleRunTest)

Information about the uploaded test to be run against the device pool.

getDevicePoolCompatibility_appArn :: Lens' GetDevicePoolCompatibility (Maybe Text)

The ARN of the app that is associated with the specified device pool.

getDevicePoolCompatibility_configuration :: Lens' GetDevicePoolCompatibility (Maybe ScheduleRunConfiguration)

An object that contains information about the settings for a run.

getDevicePoolCompatibility_testType :: Lens' GetDevicePoolCompatibility (Maybe TestType)

The test type for the specified device pool.

Allowed values include the following:

BUILTIN_FUZZ.

BUILTIN_EXPLORER. For Android, an app explorer that traverses an Android app, interacting with it and capturing screenshots at the same time.

APPIUM_JAVA_JUNIT.

APPIUM_JAVA_TESTNG.

APPIUM_PYTHON.

APPIUM_NODE.

APPIUM_RUBY.

APPIUM_WEB_JAVA_JUNIT.

APPIUM_WEB_JAVA_TESTNG.

APPIUM_WEB_PYTHON.

APPIUM_WEB_NODE.

APPIUM_WEB_RUBY.

CALABASH.

INSTRUMENTATION.

UIAUTOMATION.

UIAUTOMATOR.

XCTEST.

XCTEST_UI.

getDevicePoolCompatibility_devicePoolArn :: Lens' GetDevicePoolCompatibility Text

The device pool's ARN.

Destructuring the Response

data GetDevicePoolCompatibilityResponse

Represents the result of describe device pool compatibility request.

See: newGetDevicePoolCompatibilityResponse smart constructor.

Constructors

GetDevicePoolCompatibilityResponse'

Fields

incompatibleDevices :: Maybe [DevicePoolCompatibilityResult]

Information about incompatible devices.

compatibleDevices :: Maybe [DevicePoolCompatibilityResult]

Information about compatible devices.

httpStatus :: Int

The response's http status code.

newGetDevicePoolCompatibilityResponse

Arguments

:: Int
$sel:httpStatus:GetDevicePoolCompatibilityResponse'

-> GetDevicePoolCompatibilityResponse

Create a value of GetDevicePoolCompatibilityResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:incompatibleDevices:GetDevicePoolCompatibilityResponse', getDevicePoolCompatibilityResponse_incompatibleDevices - Information about incompatible devices.

$sel:compatibleDevices:GetDevicePoolCompatibilityResponse', getDevicePoolCompatibilityResponse_compatibleDevices - Information about compatible devices.

$sel:httpStatus:GetDevicePoolCompatibilityResponse', getDevicePoolCompatibilityResponse_httpStatus - The response's http status code.

Response Lenses

getDevicePoolCompatibilityResponse_incompatibleDevices :: Lens' GetDevicePoolCompatibilityResponse (Maybe [DevicePoolCompatibilityResult])

Information about incompatible devices.

getDevicePoolCompatibilityResponse_compatibleDevices :: Lens' GetDevicePoolCompatibilityResponse (Maybe [DevicePoolCompatibilityResult])

Information about compatible devices.

getDevicePoolCompatibilityResponse_httpStatus :: Lens' GetDevicePoolCompatibilityResponse Int

The response's http status code.