Amazonka.DeviceFarm.GetOfferingStatus

Description

Gets the current status and future status of all offerings purchased by an AWS account. The response indicates how many offerings are currently available and the offerings that will be available in the next period. The API returns a NotEligible error if the user is not permitted to invoke the operation. If you must be able to invoke this operation, contact aws-devicefarm-support@amazon.com.

This operation returns paginated results.

Creating a Request

data GetOfferingStatus

Represents the request to retrieve the offering status for the specified customer or account.

See: newGetOfferingStatus smart constructor.

Constructors

GetOfferingStatus'

Fields

nextToken :: Maybe Text

An identifier that was returned from the previous call to this operation, which can be used to return the next set of items in the list.

newGetOfferingStatus :: GetOfferingStatus

Create a value of GetOfferingStatus with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:nextToken:GetOfferingStatus', getOfferingStatus_nextToken - An identifier that was returned from the previous call to this operation, which can be used to return the next set of items in the list.

Request Lenses

getOfferingStatus_nextToken :: Lens' GetOfferingStatus (Maybe Text)

An identifier that was returned from the previous call to this operation, which can be used to return the next set of items in the list.

Destructuring the Response

data GetOfferingStatusResponse

Returns the status result for a device offering.

See: newGetOfferingStatusResponse smart constructor.

Constructors

GetOfferingStatusResponse'

Fields

nextPeriod :: Maybe (HashMap Text OfferingStatus)

When specified, gets the offering status for the next period.

current :: Maybe (HashMap Text OfferingStatus)

When specified, gets the offering status for the current period.

nextToken :: Maybe Text

An identifier that was returned from the previous call to this operation, which can be used to return the next set of items in the list.

httpStatus :: Int

The response's http status code.

newGetOfferingStatusResponse

Arguments

:: Int
$sel:httpStatus:GetOfferingStatusResponse'

-> GetOfferingStatusResponse

Create a value of GetOfferingStatusResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:nextPeriod:GetOfferingStatusResponse', getOfferingStatusResponse_nextPeriod - When specified, gets the offering status for the next period.

$sel:current:GetOfferingStatusResponse', getOfferingStatusResponse_current - When specified, gets the offering status for the current period.

$sel:nextToken:GetOfferingStatus', getOfferingStatusResponse_nextToken - An identifier that was returned from the previous call to this operation, which can be used to return the next set of items in the list.

$sel:httpStatus:GetOfferingStatusResponse', getOfferingStatusResponse_httpStatus - The response's http status code.

Response Lenses

getOfferingStatusResponse_nextPeriod :: Lens' GetOfferingStatusResponse (Maybe (HashMap Text OfferingStatus))

When specified, gets the offering status for the next period.

getOfferingStatusResponse_current :: Lens' GetOfferingStatusResponse (Maybe (HashMap Text OfferingStatus))

When specified, gets the offering status for the current period.

getOfferingStatusResponse_nextToken :: Lens' GetOfferingStatusResponse (Maybe Text)

An identifier that was returned from the previous call to this operation, which can be used to return the next set of items in the list.

getOfferingStatusResponse_httpStatus :: Lens' GetOfferingStatusResponse Int

The response's http status code.