Amazonka.DeviceFarm.ListTests

Description

Gets information about tests in a given test suite.

This operation returns paginated results.

Creating a Request

data ListTests

Represents a request to the list tests operation.

See: newListTests smart constructor.

Constructors

ListTests'

Fields

nextToken :: Maybe Text

An identifier that was returned from the previous call to this operation, which can be used to return the next set of items in the list.

arn :: Text

The test suite's Amazon Resource Name (ARN).

newListTests

Arguments

:: Text
$sel:arn:ListTests'

-> ListTests

Create a value of ListTests with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:nextToken:ListTests', listTests_nextToken - An identifier that was returned from the previous call to this operation, which can be used to return the next set of items in the list.

$sel:arn:ListTests', listTests_arn - The test suite's Amazon Resource Name (ARN).

Request Lenses

listTests_nextToken :: Lens' ListTests (Maybe Text)

An identifier that was returned from the previous call to this operation, which can be used to return the next set of items in the list.

listTests_arn :: Lens' ListTests Text

The test suite's Amazon Resource Name (ARN).

Destructuring the Response

data ListTestsResponse

Represents the result of a list tests request.

See: newListTestsResponse smart constructor.

Constructors

ListTestsResponse'

Fields

tests :: Maybe [Test]

Information about the tests.

nextToken :: Maybe Text

If the number of items that are returned is significantly large, this is an identifier that is also returned. It can be used in a subsequent call to this operation to return the next set of items in the list.

httpStatus :: Int

The response's http status code.

newListTestsResponse

Arguments

:: Int
$sel:httpStatus:ListTestsResponse'

-> ListTestsResponse

Create a value of ListTestsResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:tests:ListTestsResponse', listTestsResponse_tests - Information about the tests.

$sel:nextToken:ListTests', listTestsResponse_nextToken - If the number of items that are returned is significantly large, this is an identifier that is also returned. It can be used in a subsequent call to this operation to return the next set of items in the list.

$sel:httpStatus:ListTestsResponse', listTestsResponse_httpStatus - The response's http status code.

Response Lenses

listTestsResponse_tests :: Lens' ListTestsResponse (Maybe [Test])

Information about the tests.

listTestsResponse_nextToken :: Lens' ListTestsResponse (Maybe Text)

If the number of items that are returned is significantly large, this is an identifier that is also returned. It can be used in a subsequent call to this operation to return the next set of items in the list.

listTestsResponse_httpStatus :: Lens' ListTestsResponse Int

The response's http status code.