Amazonka.DeviceFarm.PurchaseOffering

Description

Immediately purchases offerings for an AWS account. Offerings renew with the latest total purchased quantity for an offering, unless the renewal was overridden. The API returns a NotEligible error if the user is not permitted to invoke the operation. If you must be able to invoke this operation, contact aws-devicefarm-support@amazon.com.

Creating a Request

data PurchaseOffering

Represents a request for a purchase offering.

See: newPurchaseOffering smart constructor.

Constructors

PurchaseOffering'

Fields

offeringPromotionId :: Maybe Text

The ID of the offering promotion to be applied to the purchase.

offeringId :: Text

The ID of the offering.

quantity :: Int

The number of device slots to purchase in an offering request.

newPurchaseOffering

Arguments

:: Text
$sel:offeringId:PurchaseOffering'

-> Int
$sel:quantity:PurchaseOffering'

-> PurchaseOffering

Create a value of PurchaseOffering with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:offeringPromotionId:PurchaseOffering', purchaseOffering_offeringPromotionId - The ID of the offering promotion to be applied to the purchase.

$sel:offeringId:PurchaseOffering', purchaseOffering_offeringId - The ID of the offering.

$sel:quantity:PurchaseOffering', purchaseOffering_quantity - The number of device slots to purchase in an offering request.

Request Lenses

purchaseOffering_offeringPromotionId :: Lens' PurchaseOffering (Maybe Text)

The ID of the offering promotion to be applied to the purchase.

purchaseOffering_offeringId :: Lens' PurchaseOffering Text

The ID of the offering.

purchaseOffering_quantity :: Lens' PurchaseOffering Int

The number of device slots to purchase in an offering request.

Destructuring the Response

data PurchaseOfferingResponse

The result of the purchase offering (for example, success or failure).

See: newPurchaseOfferingResponse smart constructor.

Constructors

PurchaseOfferingResponse'

Fields

offeringTransaction :: Maybe OfferingTransaction

Represents the offering transaction for the purchase result.

httpStatus :: Int

The response's http status code.

newPurchaseOfferingResponse

Arguments

:: Int
$sel:httpStatus:PurchaseOfferingResponse'

-> PurchaseOfferingResponse

Create a value of PurchaseOfferingResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:offeringTransaction:PurchaseOfferingResponse', purchaseOfferingResponse_offeringTransaction - Represents the offering transaction for the purchase result.

$sel:httpStatus:PurchaseOfferingResponse', purchaseOfferingResponse_httpStatus - The response's http status code.

Response Lenses

purchaseOfferingResponse_offeringTransaction :: Lens' PurchaseOfferingResponse (Maybe OfferingTransaction)

Represents the offering transaction for the purchase result.

purchaseOfferingResponse_httpStatus :: Lens' PurchaseOfferingResponse Int

The response's http status code.