Amazonka.DeviceFarm.Types.DevicePool

Description

Documentation

data DevicePool

Represents a collection of device types.

See: newDevicePool smart constructor.

Constructors

DevicePool'

Fields

arn :: Maybe Text

The device pool's ARN.

rules :: Maybe [Rule]

Information about the device pool's rules.

name :: Maybe Text

The device pool's name.

maxDevices :: Maybe Int

The number of devices that Device Farm can add to your device pool. Device Farm adds devices that are available and meet the criteria that you assign for the rules parameter. Depending on how many devices meet these constraints, your device pool might contain fewer devices than the value for this parameter.

By specifying the maximum number of devices, you can control the costs that you incur by running tests.

type' :: Maybe DevicePoolType

The device pool's type.

Allowed values include:

CURATED: A device pool that is created and managed by AWS Device Farm.

PRIVATE: A device pool that is created and managed by the device pool developer.

description :: Maybe Text

The device pool's description.

newDevicePool :: DevicePool

Create a value of DevicePool with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:arn:DevicePool', devicePool_arn - The device pool's ARN.

$sel:rules:DevicePool', devicePool_rules - Information about the device pool's rules.

$sel:name:DevicePool', devicePool_name - The device pool's name.

$sel:maxDevices:DevicePool', devicePool_maxDevices - The number of devices that Device Farm can add to your device pool. Device Farm adds devices that are available and meet the criteria that you assign for the rules parameter. Depending on how many devices meet these constraints, your device pool might contain fewer devices than the value for this parameter.

By specifying the maximum number of devices, you can control the costs that you incur by running tests.

$sel:type':DevicePool', devicePool_type - The device pool's type.

Allowed values include:

CURATED: A device pool that is created and managed by AWS Device Farm.

PRIVATE: A device pool that is created and managed by the device pool developer.

$sel:description:DevicePool', devicePool_description - The device pool's description.

devicePool_arn :: Lens' DevicePool (Maybe Text)

The device pool's ARN.

devicePool_rules :: Lens' DevicePool (Maybe [Rule])

Information about the device pool's rules.

devicePool_name :: Lens' DevicePool (Maybe Text)

The device pool's name.

devicePool_maxDevices :: Lens' DevicePool (Maybe Int)

The number of devices that Device Farm can add to your device pool. Device Farm adds devices that are available and meet the criteria that you assign for the rules parameter. Depending on how many devices meet these constraints, your device pool might contain fewer devices than the value for this parameter.

By specifying the maximum number of devices, you can control the costs that you incur by running tests.

devicePool_type :: Lens' DevicePool (Maybe DevicePoolType)

The device pool's type.

Allowed values include:

CURATED: A device pool that is created and managed by AWS Device Farm.

PRIVATE: A device pool that is created and managed by the device pool developer.

devicePool_description :: Lens' DevicePool (Maybe Text)

The device pool's description.