Amazonka.DeviceFarm.Types.Resolution

Description

Documentation

data Resolution

Represents the screen resolution of a device in height and width, expressed in pixels.

See: newResolution smart constructor.

Constructors

Resolution'

Fields

height :: Maybe Int

The screen resolution's height, expressed in pixels.

width :: Maybe Int

The screen resolution's width, expressed in pixels.

newResolution :: Resolution

Create a value of Resolution with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:height:Resolution', resolution_height - The screen resolution's height, expressed in pixels.

$sel:width:Resolution', resolution_width - The screen resolution's width, expressed in pixels.

resolution_height :: Lens' Resolution (Maybe Int)

The screen resolution's height, expressed in pixels.

resolution_width :: Lens' Resolution (Maybe Int)

The screen resolution's width, expressed in pixels.