Amazonka.DeviceFarm.Types.Suite

Description

Documentation

data Suite

Represents a collection of one or more tests.

See: newSuite smart constructor.

Constructors

Suite'

Fields

status :: Maybe ExecutionStatus

The suite's status.

Allowed values include:

PENDING

PENDING_CONCURRENCY

PENDING_DEVICE

PROCESSING

SCHEDULING

PREPARING

RUNNING

COMPLETED

STOPPING

counters :: Maybe Counters

The suite's result counters.

arn :: Maybe Text

The suite's ARN.

created :: Maybe POSIX

When the suite was created.

stopped :: Maybe POSIX

The suite's stop time.

result :: Maybe ExecutionResult

The suite's result.

Allowed values include:

PENDING

PASSED

WARNED

FAILED

SKIPPED

ERRORED

STOPPED

name :: Maybe Text

The suite's name.

deviceMinutes :: Maybe DeviceMinutes

Represents the total (metered or unmetered) minutes used by the test suite.

type' :: Maybe TestType

The suite's type.

Must be one of the following values:

BUILTIN_FUZZ

BUILTIN_EXPLORER

Only available for Android; an app explorer that traverses an Android app, interacting with it and capturing screenshots at the same time.

APPIUM_JAVA_JUNIT

APPIUM_JAVA_TESTNG

APPIUM_PYTHON

APPIUM_NODE

APPIUM_RUBY

APPIUM_WEB_JAVA_JUNIT

APPIUM_WEB_JAVA_TESTNG

APPIUM_WEB_PYTHON

APPIUM_WEB_NODE

APPIUM_WEB_RUBY

CALABASH

INSTRUMENTATION

UIAUTOMATION

UIAUTOMATOR

XCTEST

XCTEST_UI

message :: Maybe Text

A message about the suite's result.

started :: Maybe POSIX

The suite's start time.

newSuite :: Suite

Create a value of Suite with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:status:Suite', suite_status - The suite's status.

Allowed values include:

PENDING

PENDING_CONCURRENCY

PENDING_DEVICE

PROCESSING

SCHEDULING

PREPARING

RUNNING

COMPLETED

STOPPING

$sel:counters:Suite', suite_counters - The suite's result counters.

$sel:arn:Suite', suite_arn - The suite's ARN.

$sel:created:Suite', suite_created - When the suite was created.

$sel:stopped:Suite', suite_stopped - The suite's stop time.

$sel:result:Suite', suite_result - The suite's result.

Allowed values include:

PENDING

PASSED

WARNED

FAILED

SKIPPED

ERRORED

STOPPED

$sel:name:Suite', suite_name - The suite's name.

$sel:deviceMinutes:Suite', suite_deviceMinutes - Represents the total (metered or unmetered) minutes used by the test suite.

$sel:type':Suite', suite_type - The suite's type.

Must be one of the following values:

BUILTIN_FUZZ

BUILTIN_EXPLORER

Only available for Android; an app explorer that traverses an Android app, interacting with it and capturing screenshots at the same time.

APPIUM_JAVA_JUNIT

APPIUM_JAVA_TESTNG

APPIUM_PYTHON

APPIUM_NODE

APPIUM_RUBY

APPIUM_WEB_JAVA_JUNIT

APPIUM_WEB_JAVA_TESTNG

APPIUM_WEB_PYTHON

APPIUM_WEB_NODE

APPIUM_WEB_RUBY

CALABASH

INSTRUMENTATION

UIAUTOMATION

UIAUTOMATOR

XCTEST

XCTEST_UI

$sel:message:Suite', suite_message - A message about the suite's result.

$sel:started:Suite', suite_started - The suite's start time.

suite_status :: Lens' Suite (Maybe ExecutionStatus)

The suite's status.

Allowed values include:

PENDING

PENDING_CONCURRENCY

PENDING_DEVICE

PROCESSING

SCHEDULING

PREPARING

RUNNING

COMPLETED

STOPPING

suite_counters :: Lens' Suite (Maybe Counters)

The suite's result counters.

suite_arn :: Lens' Suite (Maybe Text)

The suite's ARN.

suite_created :: Lens' Suite (Maybe UTCTime)

When the suite was created.

suite_stopped :: Lens' Suite (Maybe UTCTime)

The suite's stop time.

suite_result :: Lens' Suite (Maybe ExecutionResult)

The suite's result.

Allowed values include:

PENDING

PASSED

WARNED

FAILED

SKIPPED

ERRORED

STOPPED

suite_name :: Lens' Suite (Maybe Text)

The suite's name.

suite_deviceMinutes :: Lens' Suite (Maybe DeviceMinutes)

Represents the total (metered or unmetered) minutes used by the test suite.

suite_type :: Lens' Suite (Maybe TestType)

The suite's type.

Must be one of the following values:

BUILTIN_FUZZ

BUILTIN_EXPLORER

Only available for Android; an app explorer that traverses an Android app, interacting with it and capturing screenshots at the same time.

APPIUM_JAVA_JUNIT

APPIUM_JAVA_TESTNG

APPIUM_PYTHON

APPIUM_NODE

APPIUM_RUBY

APPIUM_WEB_JAVA_JUNIT

APPIUM_WEB_JAVA_TESTNG

APPIUM_WEB_PYTHON

APPIUM_WEB_NODE

APPIUM_WEB_RUBY

CALABASH

INSTRUMENTATION

UIAUTOMATION

UIAUTOMATOR

XCTEST

XCTEST_UI

suite_message :: Lens' Suite (Maybe Text)

A message about the suite's result.

suite_started :: Lens' Suite (Maybe UTCTime)

The suite's start time.