Amazonka.DeviceFarm.Types.Test

Description

Documentation

data Test

Represents a condition that is evaluated.

See: newTest smart constructor.

Constructors

Test'

Fields

status :: Maybe ExecutionStatus

The test's status.

Allowed values include:

PENDING

PENDING_CONCURRENCY

PENDING_DEVICE

PROCESSING

SCHEDULING

PREPARING

RUNNING

COMPLETED

STOPPING

counters :: Maybe Counters

The test's result counters.

arn :: Maybe Text

The test's ARN.

created :: Maybe POSIX

When the test was created.

stopped :: Maybe POSIX

The test's stop time.

result :: Maybe ExecutionResult

The test's result.

Allowed values include:

PENDING

PASSED

WARNED

FAILED

SKIPPED

ERRORED

STOPPED

name :: Maybe Text

The test's name.

deviceMinutes :: Maybe DeviceMinutes

Represents the total (metered or unmetered) minutes used by the test.

type' :: Maybe TestType

The test's type.

Must be one of the following values:

BUILTIN_FUZZ

BUILTIN_EXPLORER

For Android, an app explorer that traverses an Android app, interacting with it and capturing screenshots at the same time.

APPIUM_JAVA_JUNIT

APPIUM_JAVA_TESTNG

APPIUM_PYTHON

APPIUM_NODE

APPIUM_RUBY

APPIUM_WEB_JAVA_JUNIT

APPIUM_WEB_JAVA_TESTNG

APPIUM_WEB_PYTHON

APPIUM_WEB_NODE

APPIUM_WEB_RUBY

CALABASH

INSTRUMENTATION

UIAUTOMATION

UIAUTOMATOR

XCTEST

XCTEST_UI

message :: Maybe Text

A message about the test's result.

started :: Maybe POSIX

The test's start time.

newTest :: Test

Create a value of Test with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:status:Test', test_status - The test's status.

Allowed values include:

PENDING

PENDING_CONCURRENCY

PENDING_DEVICE

PROCESSING

SCHEDULING

PREPARING

RUNNING

COMPLETED

STOPPING

$sel:counters:Test', test_counters - The test's result counters.

$sel:arn:Test', test_arn - The test's ARN.

$sel:created:Test', test_created - When the test was created.

$sel:stopped:Test', test_stopped - The test's stop time.

$sel:result:Test', test_result - The test's result.

Allowed values include:

PENDING

PASSED

WARNED

FAILED

SKIPPED

ERRORED

STOPPED

$sel:name:Test', test_name - The test's name.

$sel:deviceMinutes:Test', test_deviceMinutes - Represents the total (metered or unmetered) minutes used by the test.

$sel:type':Test', test_type - The test's type.

Must be one of the following values:

BUILTIN_FUZZ

BUILTIN_EXPLORER

For Android, an app explorer that traverses an Android app, interacting with it and capturing screenshots at the same time.

APPIUM_JAVA_JUNIT

APPIUM_JAVA_TESTNG

APPIUM_PYTHON

APPIUM_NODE

APPIUM_RUBY

APPIUM_WEB_JAVA_JUNIT

APPIUM_WEB_JAVA_TESTNG

APPIUM_WEB_PYTHON

APPIUM_WEB_NODE

APPIUM_WEB_RUBY

CALABASH

INSTRUMENTATION

UIAUTOMATION

UIAUTOMATOR

XCTEST

XCTEST_UI

$sel:message:Test', test_message - A message about the test's result.

$sel:started:Test', test_started - The test's start time.

test_status :: Lens' Test (Maybe ExecutionStatus)

The test's status.

Allowed values include:

PENDING

PENDING_CONCURRENCY

PENDING_DEVICE

PROCESSING

SCHEDULING

PREPARING

RUNNING

COMPLETED

STOPPING

test_counters :: Lens' Test (Maybe Counters)

The test's result counters.

test_arn :: Lens' Test (Maybe Text)

The test's ARN.

test_created :: Lens' Test (Maybe UTCTime)

When the test was created.

test_stopped :: Lens' Test (Maybe UTCTime)

The test's stop time.

test_result :: Lens' Test (Maybe ExecutionResult)

The test's result.

Allowed values include:

PENDING

PASSED

WARNED

FAILED

SKIPPED

ERRORED

STOPPED

test_name :: Lens' Test (Maybe Text)

The test's name.

test_deviceMinutes :: Lens' Test (Maybe DeviceMinutes)

Represents the total (metered or unmetered) minutes used by the test.

test_type :: Lens' Test (Maybe TestType)

The test's type.

Must be one of the following values:

BUILTIN_FUZZ

BUILTIN_EXPLORER

For Android, an app explorer that traverses an Android app, interacting with it and capturing screenshots at the same time.

APPIUM_JAVA_JUNIT

APPIUM_JAVA_TESTNG

APPIUM_PYTHON

APPIUM_NODE

APPIUM_RUBY

APPIUM_WEB_JAVA_JUNIT

APPIUM_WEB_JAVA_TESTNG

APPIUM_WEB_PYTHON

APPIUM_WEB_NODE

APPIUM_WEB_RUBY

CALABASH

INSTRUMENTATION

UIAUTOMATION

UIAUTOMATOR

XCTEST

XCTEST_UI

test_message :: Lens' Test (Maybe Text)

A message about the test's result.

test_started :: Lens' Test (Maybe UTCTime)

The test's start time.