Amazonka.DeviceFarm.Types.TestGridVpcConfig

Description

Documentation

data TestGridVpcConfig

The VPC security groups and subnets that are attached to a project.

See: newTestGridVpcConfig smart constructor.

Constructors

TestGridVpcConfig'

Fields

securityGroupIds :: NonEmpty Text

A list of VPC security group IDs in your Amazon VPC.

subnetIds :: NonEmpty Text

A list of VPC subnet IDs in your Amazon VPC.

vpcId :: Text

The ID of the Amazon VPC.

newTestGridVpcConfig

Arguments

:: NonEmpty Text
$sel:securityGroupIds:TestGridVpcConfig'

-> NonEmpty Text
$sel:subnetIds:TestGridVpcConfig'

-> Text
$sel:vpcId:TestGridVpcConfig'

-> TestGridVpcConfig

Create a value of TestGridVpcConfig with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:securityGroupIds:TestGridVpcConfig', testGridVpcConfig_securityGroupIds - A list of VPC security group IDs in your Amazon VPC.

$sel:subnetIds:TestGridVpcConfig', testGridVpcConfig_subnetIds - A list of VPC subnet IDs in your Amazon VPC.

$sel:vpcId:TestGridVpcConfig', testGridVpcConfig_vpcId - The ID of the Amazon VPC.

testGridVpcConfig_securityGroupIds :: Lens' TestGridVpcConfig (NonEmpty Text)

A list of VPC security group IDs in your Amazon VPC.

testGridVpcConfig_subnetIds :: Lens' TestGridVpcConfig (NonEmpty Text)

A list of VPC subnet IDs in your Amazon VPC.

testGridVpcConfig_vpcId :: Lens' TestGridVpcConfig Text

The ID of the Amazon VPC.