Amazonka.DeviceFarm.Types.UniqueProblem

Description

Documentation

data UniqueProblem

A collection of one or more problems, grouped by their result.

See: newUniqueProblem smart constructor.

Constructors

UniqueProblem'

Fields

problems :: Maybe [Problem]

Information about the problems.

message :: Maybe Text

A message about the unique problems' result.

newUniqueProblem :: UniqueProblem

Create a value of UniqueProblem with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:problems:UniqueProblem', uniqueProblem_problems - Information about the problems.

$sel:message:UniqueProblem', uniqueProblem_message - A message about the unique problems' result.

uniqueProblem_problems :: Lens' UniqueProblem (Maybe [Problem])

Information about the problems.

uniqueProblem_message :: Lens' UniqueProblem (Maybe Text)

A message about the unique problems' result.