Amazonka.DirectConnect.AssociateVirtualInterface

Description

Associates a virtual interface with a specified link aggregation group (LAG) or connection. Connectivity to Amazon Web Services is temporarily interrupted as the virtual interface is being migrated. If the target connection or LAG has an associated virtual interface with a conflicting VLAN number or a conflicting IP address, the operation fails.

Virtual interfaces associated with a hosted connection cannot be associated with a LAG; hosted connections must be migrated along with their virtual interfaces using AssociateHostedConnection.

To reassociate a virtual interface to a new connection or LAG, the requester must own either the virtual interface itself or the connection to which the virtual interface is currently associated. Additionally, the requester must own the connection or LAG for the association.

Creating a Request

data AssociateVirtualInterface

See: newAssociateVirtualInterface smart constructor.

Constructors

AssociateVirtualInterface'

Fields

virtualInterfaceId :: Text

The ID of the virtual interface.

connectionId :: Text

The ID of the LAG or connection.

newAssociateVirtualInterface

Arguments

:: Text
$sel:virtualInterfaceId:AssociateVirtualInterface'

-> Text
$sel:connectionId:AssociateVirtualInterface'

-> AssociateVirtualInterface

Create a value of AssociateVirtualInterface with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:virtualInterfaceId:AssociateVirtualInterface', associateVirtualInterface_virtualInterfaceId - The ID of the virtual interface.

$sel:connectionId:AssociateVirtualInterface', associateVirtualInterface_connectionId - The ID of the LAG or connection.

Request Lenses

associateVirtualInterface_virtualInterfaceId :: Lens' AssociateVirtualInterface Text

The ID of the virtual interface.

associateVirtualInterface_connectionId :: Lens' AssociateVirtualInterface Text

The ID of the LAG or connection.

Destructuring the Response

data VirtualInterface

Information about a virtual interface.

See: newVirtualInterface smart constructor.

Constructors

VirtualInterface'

Fields

bgpPeers :: Maybe [BGPPeer]

The BGP peers configured on this virtual interface.

virtualGatewayId :: Maybe Text

The ID of the virtual private gateway. Applies only to private virtual interfaces.

mtu :: Maybe Int

The maximum transmission unit (MTU), in bytes. The supported values are 1500 and 9001. The default value is 1500.

routeFilterPrefixes :: Maybe [RouteFilterPrefix]

The routes to be advertised to the Amazon Web Services network in this Region. Applies to public virtual interfaces.

customerAddress :: Maybe Text

The IP address assigned to the customer interface.

vlan :: Maybe Int

The ID of the VLAN.

location :: Maybe Text

The location of the connection.

amazonAddress :: Maybe Text

The IP address assigned to the Amazon interface.

addressFamily :: Maybe AddressFamily

The address family for the BGP peer.

virtualInterfaceState :: Maybe VirtualInterfaceState

The state of the virtual interface. The following are the possible values:

confirming: The creation of the virtual interface is pending confirmation from the virtual interface owner. If the owner of the virtual interface is different from the owner of the connection on which it is provisioned, then the virtual interface will remain in this state until it is confirmed by the virtual interface owner.

verifying: This state only applies to public virtual interfaces. Each public virtual interface needs validation before the virtual interface can be created.

pending: A virtual interface is in this state from the time that it is created until the virtual interface is ready to forward traffic.

available: A virtual interface that is able to forward traffic.

down: A virtual interface that is BGP down.

deleting: A virtual interface is in this state immediately after calling DeleteVirtualInterface until it can no longer forward traffic.

deleted: A virtual interface that cannot forward traffic.

rejected: The virtual interface owner has declined creation of the virtual interface. If a virtual interface in the Confirming state is deleted by the virtual interface owner, the virtual interface enters the Rejected state.

unknown: The state of the virtual interface is not available.

connectionId :: Maybe Text

The ID of the connection.

awsLogicalDeviceId :: Maybe Text

The Direct Connect endpoint that terminates the logical connection. This device might be different than the device that terminates the physical connection.

directConnectGatewayId :: Maybe Text

The ID of the Direct Connect gateway.

amazonSideAsn :: Maybe Integer

The autonomous system number (ASN) for the Amazon side of the connection.

virtualInterfaceType :: Maybe Text

The type of virtual interface. The possible values are private and public.

asn :: Maybe Int

The autonomous system (AS) number for Border Gateway Protocol (BGP) configuration.

The valid values are 1-2147483647.

authKey :: Maybe Text

The authentication key for BGP configuration. This string has a minimum length of 6 characters and and a maximun lenth of 80 characters.

jumboFrameCapable :: Maybe Bool

Indicates whether jumbo frames (9001 MTU) are supported.

customerRouterConfig :: Maybe Text

The customer router configuration.

ownerAccount :: Maybe Text

The ID of the Amazon Web Services account that owns the virtual interface.

region :: Maybe Text

The Amazon Web Services Region where the virtual interface is located.

virtualInterfaceName :: Maybe Text

The name of the virtual interface assigned by the customer network. The name has a maximum of 100 characters. The following are valid characters: a-z, 0-9 and a hyphen (-).

awsDeviceV2 :: Maybe Text

The Direct Connect endpoint that terminates the physical connection.

virtualInterfaceId :: Maybe Text

The ID of the virtual interface.

tags :: Maybe (NonEmpty Tag)

The tags associated with the virtual interface.

newVirtualInterface :: VirtualInterface

Create a value of VirtualInterface with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:bgpPeers:VirtualInterface', virtualInterface_bgpPeers - The BGP peers configured on this virtual interface.

$sel:virtualGatewayId:VirtualInterface', virtualInterface_virtualGatewayId - The ID of the virtual private gateway. Applies only to private virtual interfaces.

$sel:mtu:VirtualInterface', virtualInterface_mtu - The maximum transmission unit (MTU), in bytes. The supported values are 1500 and 9001. The default value is 1500.

$sel:routeFilterPrefixes:VirtualInterface', virtualInterface_routeFilterPrefixes - The routes to be advertised to the Amazon Web Services network in this Region. Applies to public virtual interfaces.

$sel:customerAddress:VirtualInterface', virtualInterface_customerAddress - The IP address assigned to the customer interface.

$sel:vlan:VirtualInterface', virtualInterface_vlan - The ID of the VLAN.

$sel:location:VirtualInterface', virtualInterface_location - The location of the connection.

$sel:amazonAddress:VirtualInterface', virtualInterface_amazonAddress - The IP address assigned to the Amazon interface.

$sel:addressFamily:VirtualInterface', virtualInterface_addressFamily - The address family for the BGP peer.

$sel:virtualInterfaceState:VirtualInterface', virtualInterface_virtualInterfaceState - The state of the virtual interface. The following are the possible values:

confirming: The creation of the virtual interface is pending confirmation from the virtual interface owner. If the owner of the virtual interface is different from the owner of the connection on which it is provisioned, then the virtual interface will remain in this state until it is confirmed by the virtual interface owner.

verifying: This state only applies to public virtual interfaces. Each public virtual interface needs validation before the virtual interface can be created.

pending: A virtual interface is in this state from the time that it is created until the virtual interface is ready to forward traffic.

available: A virtual interface that is able to forward traffic.

down: A virtual interface that is BGP down.

deleting: A virtual interface is in this state immediately after calling DeleteVirtualInterface until it can no longer forward traffic.

deleted: A virtual interface that cannot forward traffic.

rejected: The virtual interface owner has declined creation of the virtual interface. If a virtual interface in the Confirming state is deleted by the virtual interface owner, the virtual interface enters the Rejected state.

unknown: The state of the virtual interface is not available.

$sel:connectionId:VirtualInterface', virtualInterface_connectionId - The ID of the connection.

$sel:awsLogicalDeviceId:VirtualInterface', virtualInterface_awsLogicalDeviceId - The Direct Connect endpoint that terminates the logical connection. This device might be different than the device that terminates the physical connection.

$sel:directConnectGatewayId:VirtualInterface', virtualInterface_directConnectGatewayId - The ID of the Direct Connect gateway.

$sel:amazonSideAsn:VirtualInterface', virtualInterface_amazonSideAsn - The autonomous system number (ASN) for the Amazon side of the connection.

$sel:virtualInterfaceType:VirtualInterface', virtualInterface_virtualInterfaceType - The type of virtual interface. The possible values are private and public.

$sel:asn:VirtualInterface', virtualInterface_asn - The autonomous system (AS) number for Border Gateway Protocol (BGP) configuration.

The valid values are 1-2147483647.

$sel:authKey:VirtualInterface', virtualInterface_authKey - The authentication key for BGP configuration. This string has a minimum length of 6 characters and and a maximun lenth of 80 characters.

$sel:jumboFrameCapable:VirtualInterface', virtualInterface_jumboFrameCapable - Indicates whether jumbo frames (9001 MTU) are supported.

$sel:customerRouterConfig:VirtualInterface', virtualInterface_customerRouterConfig - The customer router configuration.

$sel:ownerAccount:VirtualInterface', virtualInterface_ownerAccount - The ID of the Amazon Web Services account that owns the virtual interface.

$sel:region:VirtualInterface', virtualInterface_region - The Amazon Web Services Region where the virtual interface is located.

$sel:virtualInterfaceName:VirtualInterface', virtualInterface_virtualInterfaceName - The name of the virtual interface assigned by the customer network. The name has a maximum of 100 characters. The following are valid characters: a-z, 0-9 and a hyphen (-).

$sel:awsDeviceV2:VirtualInterface', virtualInterface_awsDeviceV2 - The Direct Connect endpoint that terminates the physical connection.

$sel:virtualInterfaceId:VirtualInterface', virtualInterface_virtualInterfaceId - The ID of the virtual interface.

$sel:tags:VirtualInterface', virtualInterface_tags - The tags associated with the virtual interface.

Response Lenses

virtualInterface_bgpPeers :: Lens' VirtualInterface (Maybe [BGPPeer])

The BGP peers configured on this virtual interface.

virtualInterface_virtualGatewayId :: Lens' VirtualInterface (Maybe Text)

The ID of the virtual private gateway. Applies only to private virtual interfaces.

virtualInterface_mtu :: Lens' VirtualInterface (Maybe Int)

The maximum transmission unit (MTU), in bytes. The supported values are 1500 and 9001. The default value is 1500.

virtualInterface_routeFilterPrefixes :: Lens' VirtualInterface (Maybe [RouteFilterPrefix])

The routes to be advertised to the Amazon Web Services network in this Region. Applies to public virtual interfaces.

virtualInterface_customerAddress :: Lens' VirtualInterface (Maybe Text)

The IP address assigned to the customer interface.

virtualInterface_vlan :: Lens' VirtualInterface (Maybe Int)

The ID of the VLAN.

virtualInterface_location :: Lens' VirtualInterface (Maybe Text)

The location of the connection.

virtualInterface_amazonAddress :: Lens' VirtualInterface (Maybe Text)

The IP address assigned to the Amazon interface.

virtualInterface_addressFamily :: Lens' VirtualInterface (Maybe AddressFamily)

The address family for the BGP peer.

virtualInterface_virtualInterfaceState :: Lens' VirtualInterface (Maybe VirtualInterfaceState)

The state of the virtual interface. The following are the possible values:

confirming: The creation of the virtual interface is pending confirmation from the virtual interface owner. If the owner of the virtual interface is different from the owner of the connection on which it is provisioned, then the virtual interface will remain in this state until it is confirmed by the virtual interface owner.

verifying: This state only applies to public virtual interfaces. Each public virtual interface needs validation before the virtual interface can be created.

pending: A virtual interface is in this state from the time that it is created until the virtual interface is ready to forward traffic.

available: A virtual interface that is able to forward traffic.

down: A virtual interface that is BGP down.

deleting: A virtual interface is in this state immediately after calling DeleteVirtualInterface until it can no longer forward traffic.

deleted: A virtual interface that cannot forward traffic.

rejected: The virtual interface owner has declined creation of the virtual interface. If a virtual interface in the Confirming state is deleted by the virtual interface owner, the virtual interface enters the Rejected state.

unknown: The state of the virtual interface is not available.

virtualInterface_connectionId :: Lens' VirtualInterface (Maybe Text)

The ID of the connection.

virtualInterface_awsLogicalDeviceId :: Lens' VirtualInterface (Maybe Text)

The Direct Connect endpoint that terminates the logical connection. This device might be different than the device that terminates the physical connection.

virtualInterface_directConnectGatewayId :: Lens' VirtualInterface (Maybe Text)

The ID of the Direct Connect gateway.

virtualInterface_amazonSideAsn :: Lens' VirtualInterface (Maybe Integer)

The autonomous system number (ASN) for the Amazon side of the connection.

virtualInterface_virtualInterfaceType :: Lens' VirtualInterface (Maybe Text)

The type of virtual interface. The possible values are private and public.

virtualInterface_asn :: Lens' VirtualInterface (Maybe Int)

The autonomous system (AS) number for Border Gateway Protocol (BGP) configuration.

The valid values are 1-2147483647.

virtualInterface_authKey :: Lens' VirtualInterface (Maybe Text)

The authentication key for BGP configuration. This string has a minimum length of 6 characters and and a maximun lenth of 80 characters.

virtualInterface_jumboFrameCapable :: Lens' VirtualInterface (Maybe Bool)

Indicates whether jumbo frames (9001 MTU) are supported.

virtualInterface_customerRouterConfig :: Lens' VirtualInterface (Maybe Text)

The customer router configuration.

virtualInterface_ownerAccount :: Lens' VirtualInterface (Maybe Text)

The ID of the Amazon Web Services account that owns the virtual interface.

virtualInterface_region :: Lens' VirtualInterface (Maybe Text)

The Amazon Web Services Region where the virtual interface is located.

virtualInterface_virtualInterfaceName :: Lens' VirtualInterface (Maybe Text)

The name of the virtual interface assigned by the customer network. The name has a maximum of 100 characters. The following are valid characters: a-z, 0-9 and a hyphen (-).

virtualInterface_awsDeviceV2 :: Lens' VirtualInterface (Maybe Text)

The Direct Connect endpoint that terminates the physical connection.

virtualInterface_virtualInterfaceId :: Lens' VirtualInterface (Maybe Text)

The ID of the virtual interface.

virtualInterface_tags :: Lens' VirtualInterface (Maybe (NonEmpty Tag))

The tags associated with the virtual interface.