Amazonka.DirectConnect.ConfirmPrivateVirtualInterface

Description

Accepts ownership of a private virtual interface created by another Amazon Web Services account.

After the virtual interface owner makes this call, the virtual interface is created and attached to the specified virtual private gateway or Direct Connect gateway, and is made available to handle traffic.

Creating a Request

data ConfirmPrivateVirtualInterface

See: newConfirmPrivateVirtualInterface smart constructor.

Constructors

ConfirmPrivateVirtualInterface'

Fields

virtualGatewayId :: Maybe Text

The ID of the virtual private gateway.

directConnectGatewayId :: Maybe Text

The ID of the Direct Connect gateway.

virtualInterfaceId :: Text

The ID of the virtual interface.

newConfirmPrivateVirtualInterface

Arguments

:: Text
$sel:virtualInterfaceId:ConfirmPrivateVirtualInterface'

-> ConfirmPrivateVirtualInterface

Create a value of ConfirmPrivateVirtualInterface with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:virtualGatewayId:ConfirmPrivateVirtualInterface', confirmPrivateVirtualInterface_virtualGatewayId - The ID of the virtual private gateway.

$sel:directConnectGatewayId:ConfirmPrivateVirtualInterface', confirmPrivateVirtualInterface_directConnectGatewayId - The ID of the Direct Connect gateway.

$sel:virtualInterfaceId:ConfirmPrivateVirtualInterface', confirmPrivateVirtualInterface_virtualInterfaceId - The ID of the virtual interface.

Request Lenses

confirmPrivateVirtualInterface_virtualGatewayId :: Lens' ConfirmPrivateVirtualInterface (Maybe Text)

The ID of the virtual private gateway.

confirmPrivateVirtualInterface_directConnectGatewayId :: Lens' ConfirmPrivateVirtualInterface (Maybe Text)

The ID of the Direct Connect gateway.

confirmPrivateVirtualInterface_virtualInterfaceId :: Lens' ConfirmPrivateVirtualInterface Text

The ID of the virtual interface.

Destructuring the Response

data ConfirmPrivateVirtualInterfaceResponse

See: newConfirmPrivateVirtualInterfaceResponse smart constructor.

Constructors

ConfirmPrivateVirtualInterfaceResponse'

Fields

virtualInterfaceState :: Maybe VirtualInterfaceState

The state of the virtual interface. The following are the possible values:

confirming: The creation of the virtual interface is pending confirmation from the virtual interface owner. If the owner of the virtual interface is different from the owner of the connection on which it is provisioned, then the virtual interface will remain in this state until it is confirmed by the virtual interface owner.

verifying: This state only applies to public virtual interfaces. Each public virtual interface needs validation before the virtual interface can be created.

pending: A virtual interface is in this state from the time that it is created until the virtual interface is ready to forward traffic.

available: A virtual interface that is able to forward traffic.

down: A virtual interface that is BGP down.

deleting: A virtual interface is in this state immediately after calling DeleteVirtualInterface until it can no longer forward traffic.

deleted: A virtual interface that cannot forward traffic.

rejected: The virtual interface owner has declined creation of the virtual interface. If a virtual interface in the Confirming state is deleted by the virtual interface owner, the virtual interface enters the Rejected state.

unknown: The state of the virtual interface is not available.

httpStatus :: Int

The response's http status code.

newConfirmPrivateVirtualInterfaceResponse

Arguments

:: Int
$sel:httpStatus:ConfirmPrivateVirtualInterfaceResponse'

-> ConfirmPrivateVirtualInterfaceResponse

Create a value of ConfirmPrivateVirtualInterfaceResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:virtualInterfaceState:ConfirmPrivateVirtualInterfaceResponse', confirmPrivateVirtualInterfaceResponse_virtualInterfaceState - The state of the virtual interface. The following are the possible values:

confirming: The creation of the virtual interface is pending confirmation from the virtual interface owner. If the owner of the virtual interface is different from the owner of the connection on which it is provisioned, then the virtual interface will remain in this state until it is confirmed by the virtual interface owner.

verifying: This state only applies to public virtual interfaces. Each public virtual interface needs validation before the virtual interface can be created.

pending: A virtual interface is in this state from the time that it is created until the virtual interface is ready to forward traffic.

available: A virtual interface that is able to forward traffic.

down: A virtual interface that is BGP down.

deleting: A virtual interface is in this state immediately after calling DeleteVirtualInterface until it can no longer forward traffic.

deleted: A virtual interface that cannot forward traffic.

rejected: The virtual interface owner has declined creation of the virtual interface. If a virtual interface in the Confirming state is deleted by the virtual interface owner, the virtual interface enters the Rejected state.

unknown: The state of the virtual interface is not available.

$sel:httpStatus:ConfirmPrivateVirtualInterfaceResponse', confirmPrivateVirtualInterfaceResponse_httpStatus - The response's http status code.

Response Lenses

confirmPrivateVirtualInterfaceResponse_virtualInterfaceState :: Lens' ConfirmPrivateVirtualInterfaceResponse (Maybe VirtualInterfaceState)

The state of the virtual interface. The following are the possible values:

confirming: The creation of the virtual interface is pending confirmation from the virtual interface owner. If the owner of the virtual interface is different from the owner of the connection on which it is provisioned, then the virtual interface will remain in this state until it is confirmed by the virtual interface owner.

verifying: This state only applies to public virtual interfaces. Each public virtual interface needs validation before the virtual interface can be created.

pending: A virtual interface is in this state from the time that it is created until the virtual interface is ready to forward traffic.

available: A virtual interface that is able to forward traffic.

down: A virtual interface that is BGP down.

deleting: A virtual interface is in this state immediately after calling DeleteVirtualInterface until it can no longer forward traffic.

deleted: A virtual interface that cannot forward traffic.

rejected: The virtual interface owner has declined creation of the virtual interface. If a virtual interface in the Confirming state is deleted by the virtual interface owner, the virtual interface enters the Rejected state.

unknown: The state of the virtual interface is not available.

confirmPrivateVirtualInterfaceResponse_httpStatus :: Lens' ConfirmPrivateVirtualInterfaceResponse Int

The response's http status code.