Amazonka.DirectConnect.CreateDirectConnectGatewayAssociation

Description

Creates an association between a Direct Connect gateway and a virtual private gateway. The virtual private gateway must be attached to a VPC and must not be associated with another Direct Connect gateway.

Creating a Request

data CreateDirectConnectGatewayAssociation

See: newCreateDirectConnectGatewayAssociation smart constructor.

Constructors

CreateDirectConnectGatewayAssociation'

Fields

virtualGatewayId :: Maybe Text

The ID of the virtual private gateway.

addAllowedPrefixesToDirectConnectGateway :: Maybe [RouteFilterPrefix]

The Amazon VPC prefixes to advertise to the Direct Connect gateway

This parameter is required when you create an association to a transit gateway.

For information about how to set the prefixes, see Allowed Prefixes in the Direct Connect User Guide.

gatewayId :: Maybe Text

The ID of the virtual private gateway or transit gateway.

directConnectGatewayId :: Text

The ID of the Direct Connect gateway.

newCreateDirectConnectGatewayAssociation

Arguments

:: Text
$sel:directConnectGatewayId:CreateDirectConnectGatewayAssociation'

-> CreateDirectConnectGatewayAssociation

Create a value of CreateDirectConnectGatewayAssociation with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:virtualGatewayId:CreateDirectConnectGatewayAssociation', createDirectConnectGatewayAssociation_virtualGatewayId - The ID of the virtual private gateway.

$sel:addAllowedPrefixesToDirectConnectGateway:CreateDirectConnectGatewayAssociation', createDirectConnectGatewayAssociation_addAllowedPrefixesToDirectConnectGateway - The Amazon VPC prefixes to advertise to the Direct Connect gateway

This parameter is required when you create an association to a transit gateway.

For information about how to set the prefixes, see Allowed Prefixes in the Direct Connect User Guide.

$sel:gatewayId:CreateDirectConnectGatewayAssociation', createDirectConnectGatewayAssociation_gatewayId - The ID of the virtual private gateway or transit gateway.

$sel:directConnectGatewayId:CreateDirectConnectGatewayAssociation', createDirectConnectGatewayAssociation_directConnectGatewayId - The ID of the Direct Connect gateway.

Request Lenses

createDirectConnectGatewayAssociation_virtualGatewayId :: Lens' CreateDirectConnectGatewayAssociation (Maybe Text)

The ID of the virtual private gateway.

createDirectConnectGatewayAssociation_addAllowedPrefixesToDirectConnectGateway :: Lens' CreateDirectConnectGatewayAssociation (Maybe [RouteFilterPrefix])

The Amazon VPC prefixes to advertise to the Direct Connect gateway

This parameter is required when you create an association to a transit gateway.

For information about how to set the prefixes, see Allowed Prefixes in the Direct Connect User Guide.

createDirectConnectGatewayAssociation_gatewayId :: Lens' CreateDirectConnectGatewayAssociation (Maybe Text)

The ID of the virtual private gateway or transit gateway.

createDirectConnectGatewayAssociation_directConnectGatewayId :: Lens' CreateDirectConnectGatewayAssociation Text

The ID of the Direct Connect gateway.

Destructuring the Response

data CreateDirectConnectGatewayAssociationResponse

See: newCreateDirectConnectGatewayAssociationResponse smart constructor.

Constructors

CreateDirectConnectGatewayAssociationResponse'

Fields

directConnectGatewayAssociation :: Maybe DirectConnectGatewayAssociation

The association to be created.

httpStatus :: Int

The response's http status code.

newCreateDirectConnectGatewayAssociationResponse

Arguments

:: Int
$sel:httpStatus:CreateDirectConnectGatewayAssociationResponse'

-> CreateDirectConnectGatewayAssociationResponse

Create a value of CreateDirectConnectGatewayAssociationResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:directConnectGatewayAssociation:CreateDirectConnectGatewayAssociationResponse', createDirectConnectGatewayAssociationResponse_directConnectGatewayAssociation - The association to be created.

$sel:httpStatus:CreateDirectConnectGatewayAssociationResponse', createDirectConnectGatewayAssociationResponse_httpStatus - The response's http status code.

Response Lenses

createDirectConnectGatewayAssociationResponse_directConnectGatewayAssociation :: Lens' CreateDirectConnectGatewayAssociationResponse (Maybe DirectConnectGatewayAssociation)

The association to be created.

createDirectConnectGatewayAssociationResponse_httpStatus :: Lens' CreateDirectConnectGatewayAssociationResponse Int

The response's http status code.