Amazonka.DirectConnect.CreateDirectConnectGatewayAssociationProposal

Description

Creates a proposal to associate the specified virtual private gateway or transit gateway with the specified Direct Connect gateway.

You can associate a Direct Connect gateway and virtual private gateway or transit gateway that is owned by any Amazon Web Services account.

Creating a Request

data CreateDirectConnectGatewayAssociationProposal

See: newCreateDirectConnectGatewayAssociationProposal smart constructor.

Constructors

CreateDirectConnectGatewayAssociationProposal'

Fields

addAllowedPrefixesToDirectConnectGateway :: Maybe [RouteFilterPrefix]

The Amazon VPC prefixes to advertise to the Direct Connect gateway.

removeAllowedPrefixesToDirectConnectGateway :: Maybe [RouteFilterPrefix]

The Amazon VPC prefixes to no longer advertise to the Direct Connect gateway.

directConnectGatewayId :: Text

The ID of the Direct Connect gateway.

directConnectGatewayOwnerAccount :: Text

The ID of the Amazon Web Services account that owns the Direct Connect gateway.

gatewayId :: Text

The ID of the virtual private gateway or transit gateway.

newCreateDirectConnectGatewayAssociationProposal

Arguments

:: Text
$sel:directConnectGatewayId:CreateDirectConnectGatewayAssociationProposal'

-> Text
$sel:directConnectGatewayOwnerAccount:CreateDirectConnectGatewayAssociationProposal'

-> Text
$sel:gatewayId:CreateDirectConnectGatewayAssociationProposal'

-> CreateDirectConnectGatewayAssociationProposal

Create a value of CreateDirectConnectGatewayAssociationProposal with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:addAllowedPrefixesToDirectConnectGateway:CreateDirectConnectGatewayAssociationProposal', createDirectConnectGatewayAssociationProposal_addAllowedPrefixesToDirectConnectGateway - The Amazon VPC prefixes to advertise to the Direct Connect gateway.

$sel:removeAllowedPrefixesToDirectConnectGateway:CreateDirectConnectGatewayAssociationProposal', createDirectConnectGatewayAssociationProposal_removeAllowedPrefixesToDirectConnectGateway - The Amazon VPC prefixes to no longer advertise to the Direct Connect gateway.

$sel:directConnectGatewayId:CreateDirectConnectGatewayAssociationProposal', createDirectConnectGatewayAssociationProposal_directConnectGatewayId - The ID of the Direct Connect gateway.

$sel:directConnectGatewayOwnerAccount:CreateDirectConnectGatewayAssociationProposal', createDirectConnectGatewayAssociationProposal_directConnectGatewayOwnerAccount - The ID of the Amazon Web Services account that owns the Direct Connect gateway.

$sel:gatewayId:CreateDirectConnectGatewayAssociationProposal', createDirectConnectGatewayAssociationProposal_gatewayId - The ID of the virtual private gateway or transit gateway.

Request Lenses

createDirectConnectGatewayAssociationProposal_addAllowedPrefixesToDirectConnectGateway :: Lens' CreateDirectConnectGatewayAssociationProposal (Maybe [RouteFilterPrefix])

The Amazon VPC prefixes to advertise to the Direct Connect gateway.

createDirectConnectGatewayAssociationProposal_removeAllowedPrefixesToDirectConnectGateway :: Lens' CreateDirectConnectGatewayAssociationProposal (Maybe [RouteFilterPrefix])

The Amazon VPC prefixes to no longer advertise to the Direct Connect gateway.

createDirectConnectGatewayAssociationProposal_directConnectGatewayId :: Lens' CreateDirectConnectGatewayAssociationProposal Text

The ID of the Direct Connect gateway.

createDirectConnectGatewayAssociationProposal_directConnectGatewayOwnerAccount :: Lens' CreateDirectConnectGatewayAssociationProposal Text

The ID of the Amazon Web Services account that owns the Direct Connect gateway.

createDirectConnectGatewayAssociationProposal_gatewayId :: Lens' CreateDirectConnectGatewayAssociationProposal Text

The ID of the virtual private gateway or transit gateway.

Destructuring the Response

data CreateDirectConnectGatewayAssociationProposalResponse

See: newCreateDirectConnectGatewayAssociationProposalResponse smart constructor.

Constructors

CreateDirectConnectGatewayAssociationProposalResponse'

Fields

directConnectGatewayAssociationProposal :: Maybe DirectConnectGatewayAssociationProposal

Information about the Direct Connect gateway proposal.

httpStatus :: Int

The response's http status code.

newCreateDirectConnectGatewayAssociationProposalResponse

Arguments

:: Int
$sel:httpStatus:CreateDirectConnectGatewayAssociationProposalResponse'

-> CreateDirectConnectGatewayAssociationProposalResponse

Create a value of CreateDirectConnectGatewayAssociationProposalResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:directConnectGatewayAssociationProposal:CreateDirectConnectGatewayAssociationProposalResponse', createDirectConnectGatewayAssociationProposalResponse_directConnectGatewayAssociationProposal - Information about the Direct Connect gateway proposal.

$sel:httpStatus:CreateDirectConnectGatewayAssociationProposalResponse', createDirectConnectGatewayAssociationProposalResponse_httpStatus - The response's http status code.

Response Lenses

createDirectConnectGatewayAssociationProposalResponse_directConnectGatewayAssociationProposal :: Lens' CreateDirectConnectGatewayAssociationProposalResponse (Maybe DirectConnectGatewayAssociationProposal)

Information about the Direct Connect gateway proposal.

createDirectConnectGatewayAssociationProposalResponse_httpStatus :: Lens' CreateDirectConnectGatewayAssociationProposalResponse Int

The response's http status code.