Amazonka.DirectConnect.CreateInterconnect

Description

Creates an interconnect between an Direct Connect Partner's network and a specific Direct Connect location.

An interconnect is a connection that is capable of hosting other connections. The Direct Connect Partner can use an interconnect to provide Direct Connect hosted connections to customers through their own network services. Like a standard connection, an interconnect links the partner's network to an Direct Connect location over a standard Ethernet fiber-optic cable. One end is connected to the partner's router, the other to an Direct Connect router.

You can automatically add the new interconnect to a link aggregation group (LAG) by specifying a LAG ID in the request. This ensures that the new interconnect is allocated on the same Direct Connect endpoint that hosts the specified LAG. If there are no available ports on the endpoint, the request fails and no interconnect is created.

For each end customer, the Direct Connect Partner provisions a connection on their interconnect by calling AllocateHostedConnection. The end customer can then connect to Amazon Web Services resources by creating a virtual interface on their connection, using the VLAN assigned to them by the Direct Connect Partner.

Intended for use by Direct Connect Partners only.

Creating a Request

data CreateInterconnect

See: newCreateInterconnect smart constructor.

Constructors

CreateInterconnect'

Fields

lagId :: Maybe Text

The ID of the LAG.

providerName :: Maybe Text

The name of the service provider associated with the interconnect.

tags :: Maybe (NonEmpty Tag)

The tags to associate with the interconnect.

interconnectName :: Text

The name of the interconnect.

bandwidth :: Text

The port bandwidth, in Gbps. The possible values are 1 and 10.

location :: Text

The location of the interconnect.

newCreateInterconnect

Arguments

:: Text
$sel:interconnectName:CreateInterconnect'

-> Text
$sel:bandwidth:CreateInterconnect'

-> Text
$sel:location:CreateInterconnect'

-> CreateInterconnect

Create a value of CreateInterconnect with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:lagId:CreateInterconnect', createInterconnect_lagId - The ID of the LAG.

$sel:providerName:CreateInterconnect', createInterconnect_providerName - The name of the service provider associated with the interconnect.

$sel:tags:CreateInterconnect', createInterconnect_tags - The tags to associate with the interconnect.

$sel:interconnectName:CreateInterconnect', createInterconnect_interconnectName - The name of the interconnect.

$sel:bandwidth:CreateInterconnect', createInterconnect_bandwidth - The port bandwidth, in Gbps. The possible values are 1 and 10.

$sel:location:CreateInterconnect', createInterconnect_location - The location of the interconnect.

Request Lenses

createInterconnect_lagId :: Lens' CreateInterconnect (Maybe Text)

The ID of the LAG.

createInterconnect_providerName :: Lens' CreateInterconnect (Maybe Text)

The name of the service provider associated with the interconnect.

createInterconnect_tags :: Lens' CreateInterconnect (Maybe (NonEmpty Tag))

The tags to associate with the interconnect.

createInterconnect_interconnectName :: Lens' CreateInterconnect Text

The name of the interconnect.

createInterconnect_bandwidth :: Lens' CreateInterconnect Text

The port bandwidth, in Gbps. The possible values are 1 and 10.

createInterconnect_location :: Lens' CreateInterconnect Text

The location of the interconnect.

Destructuring the Response

data Interconnect

Information about an interconnect.

See: newInterconnect smart constructor.

Constructors

Interconnect'

Fields

lagId :: Maybe Text

The ID of the LAG.

interconnectId :: Maybe Text

The ID of the interconnect.

location :: Maybe Text

The location of the connection.

interconnectName :: Maybe Text

The name of the interconnect.

awsDevice :: Maybe Text

The Direct Connect endpoint on which the physical connection terminates.

hasLogicalRedundancy :: Maybe HasLogicalRedundancy

Indicates whether the interconnect supports a secondary BGP in the same address family (IPv4/IPv6).

awsLogicalDeviceId :: Maybe Text

The Direct Connect endpoint that terminates the logical connection. This device might be different than the device that terminates the physical connection.

loaIssueTime :: Maybe POSIX

The time of the most recent call to DescribeLoa for this connection.

bandwidth :: Maybe Text

The bandwidth of the connection.

jumboFrameCapable :: Maybe Bool

Indicates whether jumbo frames (9001 MTU) are supported.

interconnectState :: Maybe InterconnectState

The state of the interconnect. The following are the possible values:

requested: The initial state of an interconnect. The interconnect stays in the requested state until the Letter of Authorization (LOA) is sent to the customer.

pending: The interconnect is approved, and is being initialized.

available: The network link is up, and the interconnect is ready for use.

down: The network link is down.

deleting: The interconnect is being deleted.

deleted: The interconnect is deleted.

unknown: The state of the interconnect is not available.

region :: Maybe Text

The Amazon Web Services Region where the connection is located.

providerName :: Maybe Text

The name of the service provider associated with the interconnect.

awsDeviceV2 :: Maybe Text

The Direct Connect endpoint that terminates the physical connection.

tags :: Maybe (NonEmpty Tag)

The tags associated with the interconnect.

newInterconnect :: Interconnect

Create a value of Interconnect with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:lagId:Interconnect', interconnect_lagId - The ID of the LAG.

$sel:interconnectId:Interconnect', interconnect_interconnectId - The ID of the interconnect.

$sel:location:Interconnect', interconnect_location - The location of the connection.

$sel:interconnectName:Interconnect', interconnect_interconnectName - The name of the interconnect.

$sel:awsDevice:Interconnect', interconnect_awsDevice - The Direct Connect endpoint on which the physical connection terminates.

$sel:hasLogicalRedundancy:Interconnect', interconnect_hasLogicalRedundancy - Indicates whether the interconnect supports a secondary BGP in the same address family (IPv4/IPv6).

$sel:awsLogicalDeviceId:Interconnect', interconnect_awsLogicalDeviceId - The Direct Connect endpoint that terminates the logical connection. This device might be different than the device that terminates the physical connection.

$sel:loaIssueTime:Interconnect', interconnect_loaIssueTime - The time of the most recent call to DescribeLoa for this connection.

$sel:bandwidth:Interconnect', interconnect_bandwidth - The bandwidth of the connection.

$sel:jumboFrameCapable:Interconnect', interconnect_jumboFrameCapable - Indicates whether jumbo frames (9001 MTU) are supported.

$sel:interconnectState:Interconnect', interconnect_interconnectState - The state of the interconnect. The following are the possible values:

requested: The initial state of an interconnect. The interconnect stays in the requested state until the Letter of Authorization (LOA) is sent to the customer.

pending: The interconnect is approved, and is being initialized.

available: The network link is up, and the interconnect is ready for use.

down: The network link is down.

deleting: The interconnect is being deleted.

deleted: The interconnect is deleted.

unknown: The state of the interconnect is not available.

$sel:region:Interconnect', interconnect_region - The Amazon Web Services Region where the connection is located.

$sel:providerName:Interconnect', interconnect_providerName - The name of the service provider associated with the interconnect.

$sel:awsDeviceV2:Interconnect', interconnect_awsDeviceV2 - The Direct Connect endpoint that terminates the physical connection.

$sel:tags:Interconnect', interconnect_tags - The tags associated with the interconnect.

Response Lenses

interconnect_lagId :: Lens' Interconnect (Maybe Text)

The ID of the LAG.

interconnect_interconnectId :: Lens' Interconnect (Maybe Text)

The ID of the interconnect.

interconnect_location :: Lens' Interconnect (Maybe Text)

The location of the connection.

interconnect_interconnectName :: Lens' Interconnect (Maybe Text)

The name of the interconnect.

interconnect_awsDevice :: Lens' Interconnect (Maybe Text)

The Direct Connect endpoint on which the physical connection terminates.

interconnect_hasLogicalRedundancy :: Lens' Interconnect (Maybe HasLogicalRedundancy)

Indicates whether the interconnect supports a secondary BGP in the same address family (IPv4/IPv6).

interconnect_awsLogicalDeviceId :: Lens' Interconnect (Maybe Text)

The Direct Connect endpoint that terminates the logical connection. This device might be different than the device that terminates the physical connection.

interconnect_loaIssueTime :: Lens' Interconnect (Maybe UTCTime)

The time of the most recent call to DescribeLoa for this connection.

interconnect_bandwidth :: Lens' Interconnect (Maybe Text)

The bandwidth of the connection.

interconnect_jumboFrameCapable :: Lens' Interconnect (Maybe Bool)

Indicates whether jumbo frames (9001 MTU) are supported.

interconnect_interconnectState :: Lens' Interconnect (Maybe InterconnectState)

The state of the interconnect. The following are the possible values:

requested: The initial state of an interconnect. The interconnect stays in the requested state until the Letter of Authorization (LOA) is sent to the customer.

pending: The interconnect is approved, and is being initialized.

available: The network link is up, and the interconnect is ready for use.

down: The network link is down.

deleting: The interconnect is being deleted.

deleted: The interconnect is deleted.

unknown: The state of the interconnect is not available.

interconnect_region :: Lens' Interconnect (Maybe Text)

The Amazon Web Services Region where the connection is located.

interconnect_providerName :: Lens' Interconnect (Maybe Text)

The name of the service provider associated with the interconnect.

interconnect_awsDeviceV2 :: Lens' Interconnect (Maybe Text)

The Direct Connect endpoint that terminates the physical connection.

interconnect_tags :: Lens' Interconnect (Maybe (NonEmpty Tag))

The tags associated with the interconnect.