Amazonka.DirectConnect.DeleteDirectConnectGatewayAssociationProposal

Description

Deletes the association proposal request between the specified Direct Connect gateway and virtual private gateway or transit gateway.

Creating a Request

data DeleteDirectConnectGatewayAssociationProposal

See: newDeleteDirectConnectGatewayAssociationProposal smart constructor.

Constructors

DeleteDirectConnectGatewayAssociationProposal'

Fields

proposalId :: Text

The ID of the proposal.

newDeleteDirectConnectGatewayAssociationProposal

Arguments

:: Text
$sel:proposalId:DeleteDirectConnectGatewayAssociationProposal'

-> DeleteDirectConnectGatewayAssociationProposal

Create a value of DeleteDirectConnectGatewayAssociationProposal with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:proposalId:DeleteDirectConnectGatewayAssociationProposal', deleteDirectConnectGatewayAssociationProposal_proposalId - The ID of the proposal.

Request Lenses

deleteDirectConnectGatewayAssociationProposal_proposalId :: Lens' DeleteDirectConnectGatewayAssociationProposal Text

The ID of the proposal.

Destructuring the Response

data DeleteDirectConnectGatewayAssociationProposalResponse

See: newDeleteDirectConnectGatewayAssociationProposalResponse smart constructor.

Constructors

DeleteDirectConnectGatewayAssociationProposalResponse'

Fields

directConnectGatewayAssociationProposal :: Maybe DirectConnectGatewayAssociationProposal

The ID of the associated gateway.

httpStatus :: Int

The response's http status code.

newDeleteDirectConnectGatewayAssociationProposalResponse

Arguments

:: Int
$sel:httpStatus:DeleteDirectConnectGatewayAssociationProposalResponse'

-> DeleteDirectConnectGatewayAssociationProposalResponse

Create a value of DeleteDirectConnectGatewayAssociationProposalResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:directConnectGatewayAssociationProposal:DeleteDirectConnectGatewayAssociationProposalResponse', deleteDirectConnectGatewayAssociationProposalResponse_directConnectGatewayAssociationProposal - The ID of the associated gateway.

$sel:httpStatus:DeleteDirectConnectGatewayAssociationProposalResponse', deleteDirectConnectGatewayAssociationProposalResponse_httpStatus - The response's http status code.

Response Lenses

deleteDirectConnectGatewayAssociationProposalResponse_directConnectGatewayAssociationProposal :: Lens' DeleteDirectConnectGatewayAssociationProposalResponse (Maybe DirectConnectGatewayAssociationProposal)

The ID of the associated gateway.

deleteDirectConnectGatewayAssociationProposalResponse_httpStatus :: Lens' DeleteDirectConnectGatewayAssociationProposalResponse Int

The response's http status code.