Amazonka.DirectConnect.DeleteVirtualInterface

Description

Deletes a virtual interface.

Creating a Request

data DeleteVirtualInterface

See: newDeleteVirtualInterface smart constructor.

Constructors

DeleteVirtualInterface'

Fields

virtualInterfaceId :: Text

The ID of the virtual interface.

newDeleteVirtualInterface

Arguments

:: Text
$sel:virtualInterfaceId:DeleteVirtualInterface'

-> DeleteVirtualInterface

Create a value of DeleteVirtualInterface with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:virtualInterfaceId:DeleteVirtualInterface', deleteVirtualInterface_virtualInterfaceId - The ID of the virtual interface.

Request Lenses

deleteVirtualInterface_virtualInterfaceId :: Lens' DeleteVirtualInterface Text

The ID of the virtual interface.

Destructuring the Response

data DeleteVirtualInterfaceResponse

See: newDeleteVirtualInterfaceResponse smart constructor.

Constructors

DeleteVirtualInterfaceResponse'

Fields

virtualInterfaceState :: Maybe VirtualInterfaceState

The state of the virtual interface. The following are the possible values:

confirming: The creation of the virtual interface is pending confirmation from the virtual interface owner. If the owner of the virtual interface is different from the owner of the connection on which it is provisioned, then the virtual interface will remain in this state until it is confirmed by the virtual interface owner.

verifying: This state only applies to public virtual interfaces. Each public virtual interface needs validation before the virtual interface can be created.

pending: A virtual interface is in this state from the time that it is created until the virtual interface is ready to forward traffic.

available: A virtual interface that is able to forward traffic.

down: A virtual interface that is BGP down.

deleting: A virtual interface is in this state immediately after calling DeleteVirtualInterface until it can no longer forward traffic.

deleted: A virtual interface that cannot forward traffic.

rejected: The virtual interface owner has declined creation of the virtual interface. If a virtual interface in the Confirming state is deleted by the virtual interface owner, the virtual interface enters the Rejected state.

unknown: The state of the virtual interface is not available.

httpStatus :: Int

The response's http status code.

newDeleteVirtualInterfaceResponse

Arguments

:: Int
$sel:httpStatus:DeleteVirtualInterfaceResponse'

-> DeleteVirtualInterfaceResponse

Create a value of DeleteVirtualInterfaceResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:virtualInterfaceState:DeleteVirtualInterfaceResponse', deleteVirtualInterfaceResponse_virtualInterfaceState - The state of the virtual interface. The following are the possible values:

confirming: The creation of the virtual interface is pending confirmation from the virtual interface owner. If the owner of the virtual interface is different from the owner of the connection on which it is provisioned, then the virtual interface will remain in this state until it is confirmed by the virtual interface owner.

verifying: This state only applies to public virtual interfaces. Each public virtual interface needs validation before the virtual interface can be created.

pending: A virtual interface is in this state from the time that it is created until the virtual interface is ready to forward traffic.

available: A virtual interface that is able to forward traffic.

down: A virtual interface that is BGP down.

deleting: A virtual interface is in this state immediately after calling DeleteVirtualInterface until it can no longer forward traffic.

deleted: A virtual interface that cannot forward traffic.

rejected: The virtual interface owner has declined creation of the virtual interface. If a virtual interface in the Confirming state is deleted by the virtual interface owner, the virtual interface enters the Rejected state.

unknown: The state of the virtual interface is not available.

$sel:httpStatus:DeleteVirtualInterfaceResponse', deleteVirtualInterfaceResponse_httpStatus - The response's http status code.

Response Lenses

deleteVirtualInterfaceResponse_virtualInterfaceState :: Lens' DeleteVirtualInterfaceResponse (Maybe VirtualInterfaceState)

The state of the virtual interface. The following are the possible values:

confirming: The creation of the virtual interface is pending confirmation from the virtual interface owner. If the owner of the virtual interface is different from the owner of the connection on which it is provisioned, then the virtual interface will remain in this state until it is confirmed by the virtual interface owner.

verifying: This state only applies to public virtual interfaces. Each public virtual interface needs validation before the virtual interface can be created.

pending: A virtual interface is in this state from the time that it is created until the virtual interface is ready to forward traffic.

available: A virtual interface that is able to forward traffic.

down: A virtual interface that is BGP down.

deleting: A virtual interface is in this state immediately after calling DeleteVirtualInterface until it can no longer forward traffic.

deleted: A virtual interface that cannot forward traffic.

rejected: The virtual interface owner has declined creation of the virtual interface. If a virtual interface in the Confirming state is deleted by the virtual interface owner, the virtual interface enters the Rejected state.

unknown: The state of the virtual interface is not available.

deleteVirtualInterfaceResponse_httpStatus :: Lens' DeleteVirtualInterfaceResponse Int

The response's http status code.