Amazonka.DirectConnect.DescribeLoa

Description

Gets the LOA-CFA for a connection, interconnect, or link aggregation group (LAG).

The Letter of Authorization - Connecting Facility Assignment (LOA-CFA) is a document that is used when establishing your cross connect to Amazon Web Services at the colocation facility. For more information, see Requesting Cross Connects at Direct Connect Locations in the Direct Connect User Guide.

Creating a Request

data DescribeLoa

See: newDescribeLoa smart constructor.

Constructors

DescribeLoa'

Fields

loaContentType :: Maybe LoaContentType

The standard media type for the LOA-CFA document. The only supported value is application/pdf.

providerName :: Maybe Text

The name of the service provider who establishes connectivity on your behalf. If you specify this parameter, the LOA-CFA lists the provider name alongside your company name as the requester of the cross connect.

connectionId :: Text

The ID of a connection, LAG, or interconnect.

newDescribeLoa

Arguments

:: Text
$sel:connectionId:DescribeLoa'

-> DescribeLoa

Create a value of DescribeLoa with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:loaContentType:DescribeLoa', describeLoa_loaContentType - The standard media type for the LOA-CFA document. The only supported value is application/pdf.

$sel:providerName:DescribeLoa', describeLoa_providerName - The name of the service provider who establishes connectivity on your behalf. If you specify this parameter, the LOA-CFA lists the provider name alongside your company name as the requester of the cross connect.

$sel:connectionId:DescribeLoa', describeLoa_connectionId - The ID of a connection, LAG, or interconnect.

Request Lenses

describeLoa_loaContentType :: Lens' DescribeLoa (Maybe LoaContentType)

The standard media type for the LOA-CFA document. The only supported value is application/pdf.

describeLoa_providerName :: Lens' DescribeLoa (Maybe Text)

The name of the service provider who establishes connectivity on your behalf. If you specify this parameter, the LOA-CFA lists the provider name alongside your company name as the requester of the cross connect.

describeLoa_connectionId :: Lens' DescribeLoa Text

The ID of a connection, LAG, or interconnect.

Destructuring the Response

data DescribeLoaResponse

Information about a Letter of Authorization - Connecting Facility Assignment (LOA-CFA) for a connection.

See: newDescribeLoaResponse smart constructor.

Constructors

DescribeLoaResponse'

Fields

loaContent :: Maybe Base64

The binary contents of the LOA-CFA document.

loaContentType :: Maybe LoaContentType

The standard media type for the LOA-CFA document. The only supported value is application/pdf.

httpStatus :: Int

The response's http status code.

newDescribeLoaResponse

Arguments

:: Int
$sel:httpStatus:DescribeLoaResponse'

-> DescribeLoaResponse

Create a value of DescribeLoaResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:loaContent:DescribeLoaResponse', describeLoaResponse_loaContent - The binary contents of the LOA-CFA document.-- -- Note: This Lens automatically encodes and decodes Base64 data. -- The underlying isomorphism will encode to Base64 representation during -- serialisation, and decode from Base64 representation during deserialisation. -- This Lens accepts and returns only raw unencoded data.

$sel:loaContentType:DescribeLoa', describeLoaResponse_loaContentType - The standard media type for the LOA-CFA document. The only supported value is application/pdf.

$sel:httpStatus:DescribeLoaResponse', describeLoaResponse_httpStatus - The response's http status code.

Response Lenses

describeLoaResponse_loaContent :: Lens' DescribeLoaResponse (Maybe ByteString)

The binary contents of the LOA-CFA document.-- -- Note: This Lens automatically encodes and decodes Base64 data. -- The underlying isomorphism will encode to Base64 representation during -- serialisation, and decode from Base64 representation during deserialisation. -- This Lens accepts and returns only raw unencoded data.

describeLoaResponse_loaContentType :: Lens' DescribeLoaResponse (Maybe LoaContentType)

The standard media type for the LOA-CFA document. The only supported value is application/pdf.

describeLoaResponse_httpStatus :: Lens' DescribeLoaResponse Int

The response's http status code.