Amazonka.DirectConnect.StartBgpFailoverTest

Description

Starts the virtual interface failover test that verifies your configuration meets your resiliency requirements by placing the BGP peering session in the DOWN state. You can then send traffic to verify that there are no outages.

You can run the test on public, private, transit, and hosted virtual interfaces.

You can use ListVirtualInterfaceTestHistory to view the virtual interface test history.

If you need to stop the test before the test interval completes, use StopBgpFailoverTest.

Creating a Request

data StartBgpFailoverTest

See: newStartBgpFailoverTest smart constructor.

Constructors

StartBgpFailoverTest'

Fields

bgpPeers :: Maybe [Text]

The BGP peers to place in the DOWN state.

testDurationInMinutes :: Maybe Int

The time in minutes that the virtual interface failover test will last.

Maximum value: 180 minutes (3 hours).

Default: 180 minutes (3 hours).

virtualInterfaceId :: Text

The ID of the virtual interface you want to test.

newStartBgpFailoverTest

Arguments

:: Text
$sel:virtualInterfaceId:StartBgpFailoverTest'

-> StartBgpFailoverTest

Create a value of StartBgpFailoverTest with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:bgpPeers:StartBgpFailoverTest', startBgpFailoverTest_bgpPeers - The BGP peers to place in the DOWN state.

$sel:testDurationInMinutes:StartBgpFailoverTest', startBgpFailoverTest_testDurationInMinutes - The time in minutes that the virtual interface failover test will last.

Maximum value: 180 minutes (3 hours).

Default: 180 minutes (3 hours).

$sel:virtualInterfaceId:StartBgpFailoverTest', startBgpFailoverTest_virtualInterfaceId - The ID of the virtual interface you want to test.

Request Lenses

startBgpFailoverTest_bgpPeers :: Lens' StartBgpFailoverTest (Maybe [Text])

The BGP peers to place in the DOWN state.

startBgpFailoverTest_testDurationInMinutes :: Lens' StartBgpFailoverTest (Maybe Int)

The time in minutes that the virtual interface failover test will last.

Maximum value: 180 minutes (3 hours).

Default: 180 minutes (3 hours).

startBgpFailoverTest_virtualInterfaceId :: Lens' StartBgpFailoverTest Text

The ID of the virtual interface you want to test.

Destructuring the Response

data StartBgpFailoverTestResponse

See: newStartBgpFailoverTestResponse smart constructor.

Constructors

StartBgpFailoverTestResponse'

Fields

virtualInterfaceTest :: Maybe VirtualInterfaceTestHistory

Information about the virtual interface failover test.

httpStatus :: Int

The response's http status code.

newStartBgpFailoverTestResponse

Arguments

:: Int
$sel:httpStatus:StartBgpFailoverTestResponse'

-> StartBgpFailoverTestResponse

Create a value of StartBgpFailoverTestResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:virtualInterfaceTest:StartBgpFailoverTestResponse', startBgpFailoverTestResponse_virtualInterfaceTest - Information about the virtual interface failover test.

$sel:httpStatus:StartBgpFailoverTestResponse', startBgpFailoverTestResponse_httpStatus - The response's http status code.

Response Lenses

startBgpFailoverTestResponse_virtualInterfaceTest :: Lens' StartBgpFailoverTestResponse (Maybe VirtualInterfaceTestHistory)

Information about the virtual interface failover test.

startBgpFailoverTestResponse_httpStatus :: Lens' StartBgpFailoverTestResponse Int

The response's http status code.