Amazonka.DirectConnect.Types.BGPPeer

Description

Documentation

data BGPPeer

Information about a BGP peer.

See: newBGPPeer smart constructor.

Constructors

BGPPeer'

Fields

customerAddress :: Maybe Text

The IP address assigned to the customer interface.

amazonAddress :: Maybe Text

The IP address assigned to the Amazon interface.

addressFamily :: Maybe AddressFamily

The address family for the BGP peer.

awsLogicalDeviceId :: Maybe Text

The Direct Connect endpoint that terminates the logical connection. This device might be different than the device that terminates the physical connection.

bgpStatus :: Maybe BGPStatus

The status of the BGP peer. The following are the possible values:

up: The BGP peer is established. This state does not indicate the state of the routing function. Ensure that you are receiving routes over the BGP session.

down: The BGP peer is down.

unknown: The BGP peer status is not available.

asn :: Maybe Int

The autonomous system (AS) number for Border Gateway Protocol (BGP) configuration.

authKey :: Maybe Text

The authentication key for BGP configuration. This string has a minimum length of 6 characters and and a maximun lenth of 80 characters.

bgpPeerId :: Maybe Text

The ID of the BGP peer.

bgpPeerState :: Maybe BGPPeerState

The state of the BGP peer. The following are the possible values:

verifying: The BGP peering addresses or ASN require validation before the BGP peer can be created. This state applies only to public virtual interfaces.

pending: The BGP peer is created, and remains in this state until it is ready to be established.

available: The BGP peer is ready to be established.

deleting: The BGP peer is being deleted.

deleted: The BGP peer is deleted and cannot be established.

awsDeviceV2 :: Maybe Text

The Direct Connect endpoint that terminates the BGP peer.

newBGPPeer :: BGPPeer

Create a value of BGPPeer with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:customerAddress:BGPPeer', bGPPeer_customerAddress - The IP address assigned to the customer interface.

$sel:amazonAddress:BGPPeer', bGPPeer_amazonAddress - The IP address assigned to the Amazon interface.

$sel:addressFamily:BGPPeer', bGPPeer_addressFamily - The address family for the BGP peer.

$sel:awsLogicalDeviceId:BGPPeer', bGPPeer_awsLogicalDeviceId - The Direct Connect endpoint that terminates the logical connection. This device might be different than the device that terminates the physical connection.

$sel:bgpStatus:BGPPeer', bGPPeer_bgpStatus - The status of the BGP peer. The following are the possible values:

up: The BGP peer is established. This state does not indicate the state of the routing function. Ensure that you are receiving routes over the BGP session.

down: The BGP peer is down.

unknown: The BGP peer status is not available.

$sel:asn:BGPPeer', bGPPeer_asn - The autonomous system (AS) number for Border Gateway Protocol (BGP) configuration.

$sel:authKey:BGPPeer', bGPPeer_authKey - The authentication key for BGP configuration. This string has a minimum length of 6 characters and and a maximun lenth of 80 characters.

$sel:bgpPeerId:BGPPeer', bGPPeer_bgpPeerId - The ID of the BGP peer.

$sel:bgpPeerState:BGPPeer', bGPPeer_bgpPeerState - The state of the BGP peer. The following are the possible values:

verifying: The BGP peering addresses or ASN require validation before the BGP peer can be created. This state applies only to public virtual interfaces.

pending: The BGP peer is created, and remains in this state until it is ready to be established.

available: The BGP peer is ready to be established.

deleting: The BGP peer is being deleted.

deleted: The BGP peer is deleted and cannot be established.

$sel:awsDeviceV2:BGPPeer', bGPPeer_awsDeviceV2 - The Direct Connect endpoint that terminates the BGP peer.

bGPPeer_customerAddress :: Lens' BGPPeer (Maybe Text)

The IP address assigned to the customer interface.

bGPPeer_amazonAddress :: Lens' BGPPeer (Maybe Text)

The IP address assigned to the Amazon interface.

bGPPeer_addressFamily :: Lens' BGPPeer (Maybe AddressFamily)

The address family for the BGP peer.

bGPPeer_awsLogicalDeviceId :: Lens' BGPPeer (Maybe Text)

The Direct Connect endpoint that terminates the logical connection. This device might be different than the device that terminates the physical connection.

bGPPeer_bgpStatus :: Lens' BGPPeer (Maybe BGPStatus)

The status of the BGP peer. The following are the possible values:

up: The BGP peer is established. This state does not indicate the state of the routing function. Ensure that you are receiving routes over the BGP session.

down: The BGP peer is down.

unknown: The BGP peer status is not available.

bGPPeer_asn :: Lens' BGPPeer (Maybe Int)

The autonomous system (AS) number for Border Gateway Protocol (BGP) configuration.

bGPPeer_authKey :: Lens' BGPPeer (Maybe Text)

The authentication key for BGP configuration. This string has a minimum length of 6 characters and and a maximun lenth of 80 characters.

bGPPeer_bgpPeerId :: Lens' BGPPeer (Maybe Text)

The ID of the BGP peer.

bGPPeer_bgpPeerState :: Lens' BGPPeer (Maybe BGPPeerState)

The state of the BGP peer. The following are the possible values:

verifying: The BGP peering addresses or ASN require validation before the BGP peer can be created. This state applies only to public virtual interfaces.

pending: The BGP peer is created, and remains in this state until it is ready to be established.

available: The BGP peer is ready to be established.

deleting: The BGP peer is being deleted.

deleted: The BGP peer is deleted and cannot be established.

bGPPeer_awsDeviceV2 :: Lens' BGPPeer (Maybe Text)

The Direct Connect endpoint that terminates the BGP peer.