Amazonka.DirectConnect.Types.DirectConnectGatewayAssociationProposal

Description

Documentation

data DirectConnectGatewayAssociationProposal

Information about the proposal request to attach a virtual private gateway to a Direct Connect gateway.

See: newDirectConnectGatewayAssociationProposal smart constructor.

Constructors

DirectConnectGatewayAssociationProposal'

Fields

existingAllowedPrefixesToDirectConnectGateway :: Maybe [RouteFilterPrefix]

The existing Amazon VPC prefixes advertised to the Direct Connect gateway.

directConnectGatewayId :: Maybe Text

The ID of the Direct Connect gateway.

proposalId :: Maybe Text

The ID of the association proposal.

associatedGateway :: Maybe AssociatedGateway

Information about the associated gateway.

proposalState :: Maybe DirectConnectGatewayAssociationProposalState

The state of the proposal. The following are possible values:

accepted: The proposal has been accepted. The Direct Connect gateway association is available to use in this state.

deleted: The proposal has been deleted by the owner that made the proposal. The Direct Connect gateway association cannot be used in this state.

requested: The proposal has been requested. The Direct Connect gateway association cannot be used in this state.

directConnectGatewayOwnerAccount :: Maybe Text

The ID of the Amazon Web Services account that owns the Direct Connect gateway.

requestedAllowedPrefixesToDirectConnectGateway :: Maybe [RouteFilterPrefix]

The Amazon VPC prefixes to advertise to the Direct Connect gateway.

newDirectConnectGatewayAssociationProposal :: DirectConnectGatewayAssociationProposal

Create a value of DirectConnectGatewayAssociationProposal with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:existingAllowedPrefixesToDirectConnectGateway:DirectConnectGatewayAssociationProposal', directConnectGatewayAssociationProposal_existingAllowedPrefixesToDirectConnectGateway - The existing Amazon VPC prefixes advertised to the Direct Connect gateway.

$sel:directConnectGatewayId:DirectConnectGatewayAssociationProposal', directConnectGatewayAssociationProposal_directConnectGatewayId - The ID of the Direct Connect gateway.

$sel:proposalId:DirectConnectGatewayAssociationProposal', directConnectGatewayAssociationProposal_proposalId - The ID of the association proposal.

$sel:associatedGateway:DirectConnectGatewayAssociationProposal', directConnectGatewayAssociationProposal_associatedGateway - Information about the associated gateway.

$sel:proposalState:DirectConnectGatewayAssociationProposal', directConnectGatewayAssociationProposal_proposalState - The state of the proposal. The following are possible values:

accepted: The proposal has been accepted. The Direct Connect gateway association is available to use in this state.

deleted: The proposal has been deleted by the owner that made the proposal. The Direct Connect gateway association cannot be used in this state.

requested: The proposal has been requested. The Direct Connect gateway association cannot be used in this state.

$sel:directConnectGatewayOwnerAccount:DirectConnectGatewayAssociationProposal', directConnectGatewayAssociationProposal_directConnectGatewayOwnerAccount - The ID of the Amazon Web Services account that owns the Direct Connect gateway.

$sel:requestedAllowedPrefixesToDirectConnectGateway:DirectConnectGatewayAssociationProposal', directConnectGatewayAssociationProposal_requestedAllowedPrefixesToDirectConnectGateway - The Amazon VPC prefixes to advertise to the Direct Connect gateway.

directConnectGatewayAssociationProposal_existingAllowedPrefixesToDirectConnectGateway :: Lens' DirectConnectGatewayAssociationProposal (Maybe [RouteFilterPrefix])

The existing Amazon VPC prefixes advertised to the Direct Connect gateway.

directConnectGatewayAssociationProposal_directConnectGatewayId :: Lens' DirectConnectGatewayAssociationProposal (Maybe Text)

The ID of the Direct Connect gateway.

directConnectGatewayAssociationProposal_proposalId :: Lens' DirectConnectGatewayAssociationProposal (Maybe Text)

The ID of the association proposal.

directConnectGatewayAssociationProposal_associatedGateway :: Lens' DirectConnectGatewayAssociationProposal (Maybe AssociatedGateway)

Information about the associated gateway.

directConnectGatewayAssociationProposal_proposalState :: Lens' DirectConnectGatewayAssociationProposal (Maybe DirectConnectGatewayAssociationProposalState)

The state of the proposal. The following are possible values:

accepted: The proposal has been accepted. The Direct Connect gateway association is available to use in this state.

deleted: The proposal has been deleted by the owner that made the proposal. The Direct Connect gateway association cannot be used in this state.

requested: The proposal has been requested. The Direct Connect gateway association cannot be used in this state.

directConnectGatewayAssociationProposal_directConnectGatewayOwnerAccount :: Lens' DirectConnectGatewayAssociationProposal (Maybe Text)

The ID of the Amazon Web Services account that owns the Direct Connect gateway.

directConnectGatewayAssociationProposal_requestedAllowedPrefixesToDirectConnectGateway :: Lens' DirectConnectGatewayAssociationProposal (Maybe [RouteFilterPrefix])

The Amazon VPC prefixes to advertise to the Direct Connect gateway.