Amazonka.DirectConnect.Types.Location

Description

Documentation

data Location

Information about an Direct Connect location.

See: newLocation smart constructor.

Constructors

Location'

Fields

availablePortSpeeds :: Maybe [Text]

The available port speeds for the location.

locationName :: Maybe Text

The name of the location. This includes the name of the colocation partner and the physical site of the building.

locationCode :: Maybe Text

The code for the location.

region :: Maybe Text

The Amazon Web Services Region for the location.

availableProviders :: Maybe [Text]

The name of the service provider for the location.

availableMacSecPortSpeeds :: Maybe [Text]

The available MAC Security (MACsec) port speeds for the location.

newLocation :: Location

Create a value of Location with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:availablePortSpeeds:Location', location_availablePortSpeeds - The available port speeds for the location.

$sel:locationName:Location', location_locationName - The name of the location. This includes the name of the colocation partner and the physical site of the building.

$sel:locationCode:Location', location_locationCode - The code for the location.

$sel:region:Location', location_region - The Amazon Web Services Region for the location.

$sel:availableProviders:Location', location_availableProviders - The name of the service provider for the location.

$sel:availableMacSecPortSpeeds:Location', location_availableMacSecPortSpeeds - The available MAC Security (MACsec) port speeds for the location.

location_availablePortSpeeds :: Lens' Location (Maybe [Text])

The available port speeds for the location.

location_locationName :: Lens' Location (Maybe Text)

The name of the location. This includes the name of the colocation partner and the physical site of the building.

location_locationCode :: Lens' Location (Maybe Text)

The code for the location.

location_region :: Lens' Location (Maybe Text)

The Amazon Web Services Region for the location.

location_availableProviders :: Lens' Location (Maybe [Text])

The name of the service provider for the location.

location_availableMacSecPortSpeeds :: Lens' Location (Maybe [Text])

The available MAC Security (MACsec) port speeds for the location.