Amazonka.DirectConnect.Types.MacSecKey

Description

Documentation

data MacSecKey

Information about the MAC Security (MACsec) secret key.

See: newMacSecKey smart constructor.

Constructors

MacSecKey'

Fields

state :: Maybe Text

The state of the MAC Security (MACsec) secret key.

The possible values are:

associating: The MAC Security (MACsec) secret key is being validated and not yet associated with the connection or LAG.

associated: The MAC Security (MACsec) secret key is validated and associated with the connection or LAG.

disassociating: The MAC Security (MACsec) secret key is being disassociated from the connection or LAG

disassociated: The MAC Security (MACsec) secret key is no longer associated with the connection or LAG.

ckn :: Maybe Text

The Connection Key Name (CKN) for the MAC Security secret key.

secretARN :: Maybe Text

The Amazon Resource Name (ARN) of the MAC Security (MACsec) secret key.

startOn :: Maybe Text

The date that the MAC Security (MACsec) secret key takes effect. The value is displayed in UTC format.

newMacSecKey :: MacSecKey

Create a value of MacSecKey with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:state:MacSecKey', macSecKey_state - The state of the MAC Security (MACsec) secret key.

The possible values are:

associating: The MAC Security (MACsec) secret key is being validated and not yet associated with the connection or LAG.

associated: The MAC Security (MACsec) secret key is validated and associated with the connection or LAG.

disassociating: The MAC Security (MACsec) secret key is being disassociated from the connection or LAG

disassociated: The MAC Security (MACsec) secret key is no longer associated with the connection or LAG.

$sel:ckn:MacSecKey', macSecKey_ckn - The Connection Key Name (CKN) for the MAC Security secret key.

$sel:secretARN:MacSecKey', macSecKey_secretARN - The Amazon Resource Name (ARN) of the MAC Security (MACsec) secret key.

$sel:startOn:MacSecKey', macSecKey_startOn - The date that the MAC Security (MACsec) secret key takes effect. The value is displayed in UTC format.

macSecKey_state :: Lens' MacSecKey (Maybe Text)

The state of the MAC Security (MACsec) secret key.

The possible values are:

associating: The MAC Security (MACsec) secret key is being validated and not yet associated with the connection or LAG.

associated: The MAC Security (MACsec) secret key is validated and associated with the connection or LAG.

disassociating: The MAC Security (MACsec) secret key is being disassociated from the connection or LAG

disassociated: The MAC Security (MACsec) secret key is no longer associated with the connection or LAG.

macSecKey_ckn :: Lens' MacSecKey (Maybe Text)

The Connection Key Name (CKN) for the MAC Security secret key.

macSecKey_secretARN :: Lens' MacSecKey (Maybe Text)

The Amazon Resource Name (ARN) of the MAC Security (MACsec) secret key.

macSecKey_startOn :: Lens' MacSecKey (Maybe Text)

The date that the MAC Security (MACsec) secret key takes effect. The value is displayed in UTC format.