Amazonka.DirectConnect.Types.RouterType

Description

Documentation

data RouterType

Information about the virtual router.

See: newRouterType smart constructor.

Constructors

RouterType'

Fields

vendor :: Maybe Text

The vendor for the virtual interface's router.

platform :: Maybe Text

The virtual interface router platform.

xsltTemplateName :: Maybe Text

The template for the virtual interface's router.

software :: Maybe Text

The router software.

xsltTemplateNameForMacSec :: Maybe Text

The MAC Security (MACsec) template for the virtual interface's router.

routerTypeIdentifier :: Maybe Text

Identifies the router by a combination of vendor, platform, and software version. For example, CiscoSystemsInc-2900SeriesRouters-IOS124.

newRouterType :: RouterType

Create a value of RouterType with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:vendor:RouterType', routerType_vendor - The vendor for the virtual interface's router.

$sel:platform:RouterType', routerType_platform - The virtual interface router platform.

$sel:xsltTemplateName:RouterType', routerType_xsltTemplateName - The template for the virtual interface's router.

$sel:software:RouterType', routerType_software - The router software.

$sel:xsltTemplateNameForMacSec:RouterType', routerType_xsltTemplateNameForMacSec - The MAC Security (MACsec) template for the virtual interface's router.

$sel:routerTypeIdentifier:RouterType', routerType_routerTypeIdentifier - Identifies the router by a combination of vendor, platform, and software version. For example, CiscoSystemsInc-2900SeriesRouters-IOS124.

routerType_vendor :: Lens' RouterType (Maybe Text)

The vendor for the virtual interface's router.

routerType_platform :: Lens' RouterType (Maybe Text)

The virtual interface router platform.

routerType_xsltTemplateName :: Lens' RouterType (Maybe Text)

The template for the virtual interface's router.

routerType_software :: Lens' RouterType (Maybe Text)

The router software.

routerType_xsltTemplateNameForMacSec :: Lens' RouterType (Maybe Text)

The MAC Security (MACsec) template for the virtual interface's router.

routerType_routerTypeIdentifier :: Lens' RouterType (Maybe Text)

Identifies the router by a combination of vendor, platform, and software version. For example, CiscoSystemsInc-2900SeriesRouters-IOS124.