Amazonka.DirectConnect.Types.VirtualInterfaceTestHistory

Description

Documentation

data VirtualInterfaceTestHistory

Information about the virtual interface failover test.

See: newVirtualInterfaceTestHistory smart constructor.

Constructors

VirtualInterfaceTestHistory'

Fields

bgpPeers :: Maybe [Text]

The BGP peers that were put in the DOWN state as part of the virtual interface failover test.

status :: Maybe Text

The status of the virtual interface failover test.

testDurationInMinutes :: Maybe Int

The time that the virtual interface failover test ran in minutes.

startTime :: Maybe POSIX

The time that the virtual interface moves to the DOWN state.

testId :: Maybe Text

The ID of the virtual interface failover test.

endTime :: Maybe POSIX

The time that the virtual interface moves out of the DOWN state.

ownerAccount :: Maybe Text

The owner ID of the tested virtual interface.

virtualInterfaceId :: Maybe Text

The ID of the tested virtual interface.

newVirtualInterfaceTestHistory :: VirtualInterfaceTestHistory

Create a value of VirtualInterfaceTestHistory with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:bgpPeers:VirtualInterfaceTestHistory', virtualInterfaceTestHistory_bgpPeers - The BGP peers that were put in the DOWN state as part of the virtual interface failover test.

$sel:status:VirtualInterfaceTestHistory', virtualInterfaceTestHistory_status - The status of the virtual interface failover test.

$sel:testDurationInMinutes:VirtualInterfaceTestHistory', virtualInterfaceTestHistory_testDurationInMinutes - The time that the virtual interface failover test ran in minutes.

$sel:startTime:VirtualInterfaceTestHistory', virtualInterfaceTestHistory_startTime - The time that the virtual interface moves to the DOWN state.

$sel:testId:VirtualInterfaceTestHistory', virtualInterfaceTestHistory_testId - The ID of the virtual interface failover test.

$sel:endTime:VirtualInterfaceTestHistory', virtualInterfaceTestHistory_endTime - The time that the virtual interface moves out of the DOWN state.

$sel:ownerAccount:VirtualInterfaceTestHistory', virtualInterfaceTestHistory_ownerAccount - The owner ID of the tested virtual interface.

$sel:virtualInterfaceId:VirtualInterfaceTestHistory', virtualInterfaceTestHistory_virtualInterfaceId - The ID of the tested virtual interface.

virtualInterfaceTestHistory_bgpPeers :: Lens' VirtualInterfaceTestHistory (Maybe [Text])

The BGP peers that were put in the DOWN state as part of the virtual interface failover test.

virtualInterfaceTestHistory_status :: Lens' VirtualInterfaceTestHistory (Maybe Text)

The status of the virtual interface failover test.

virtualInterfaceTestHistory_testDurationInMinutes :: Lens' VirtualInterfaceTestHistory (Maybe Int)

The time that the virtual interface failover test ran in minutes.

virtualInterfaceTestHistory_startTime :: Lens' VirtualInterfaceTestHistory (Maybe UTCTime)

The time that the virtual interface moves to the DOWN state.

virtualInterfaceTestHistory_testId :: Lens' VirtualInterfaceTestHistory (Maybe Text)

The ID of the virtual interface failover test.

virtualInterfaceTestHistory_endTime :: Lens' VirtualInterfaceTestHistory (Maybe UTCTime)

The time that the virtual interface moves out of the DOWN state.

virtualInterfaceTestHistory_ownerAccount :: Lens' VirtualInterfaceTestHistory (Maybe Text)

The owner ID of the tested virtual interface.

virtualInterfaceTestHistory_virtualInterfaceId :: Lens' VirtualInterfaceTestHistory (Maybe Text)

The ID of the tested virtual interface.