Amazonka.DirectoryService.CreateComputer

Description

Creates an Active Directory computer object in the specified directory.

Creating a Request

data CreateComputer

Contains the inputs for the CreateComputer operation.

See: newCreateComputer smart constructor.

Constructors

CreateComputer'

Fields

computerAttributes :: Maybe [Attribute]

An array of Attribute objects that contain any LDAP attributes to apply to the computer account.

organizationalUnitDistinguishedName :: Maybe Text

The fully-qualified distinguished name of the organizational unit to place the computer account in.

directoryId :: Text

The identifier of the directory in which to create the computer account.

computerName :: Text

The name of the computer account.

password :: Sensitive Text

A one-time password that is used to join the computer to the directory. You should generate a random, strong password to use for this parameter.

newCreateComputer

Arguments

:: Text
$sel:directoryId:CreateComputer'

-> Text
$sel:computerName:CreateComputer'

-> Text
$sel:password:CreateComputer'

-> CreateComputer

Create a value of CreateComputer with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:computerAttributes:CreateComputer', createComputer_computerAttributes - An array of Attribute objects that contain any LDAP attributes to apply to the computer account.

$sel:organizationalUnitDistinguishedName:CreateComputer', createComputer_organizationalUnitDistinguishedName - The fully-qualified distinguished name of the organizational unit to place the computer account in.

$sel:directoryId:CreateComputer', createComputer_directoryId - The identifier of the directory in which to create the computer account.

$sel:computerName:CreateComputer', createComputer_computerName - The name of the computer account.

$sel:password:CreateComputer', createComputer_password - A one-time password that is used to join the computer to the directory. You should generate a random, strong password to use for this parameter.

Request Lenses

createComputer_computerAttributes :: Lens' CreateComputer (Maybe [Attribute])

An array of Attribute objects that contain any LDAP attributes to apply to the computer account.

createComputer_organizationalUnitDistinguishedName :: Lens' CreateComputer (Maybe Text)

The fully-qualified distinguished name of the organizational unit to place the computer account in.

createComputer_directoryId :: Lens' CreateComputer Text

The identifier of the directory in which to create the computer account.

createComputer_computerName :: Lens' CreateComputer Text

The name of the computer account.

createComputer_password :: Lens' CreateComputer Text

A one-time password that is used to join the computer to the directory. You should generate a random, strong password to use for this parameter.

Destructuring the Response

data CreateComputerResponse

Contains the results for the CreateComputer operation.

See: newCreateComputerResponse smart constructor.

Constructors

CreateComputerResponse'

Fields

computer :: Maybe Computer

A Computer object that represents the computer account.

httpStatus :: Int

The response's http status code.

newCreateComputerResponse

Arguments

:: Int
$sel:httpStatus:CreateComputerResponse'

-> CreateComputerResponse

Create a value of CreateComputerResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:computer:CreateComputerResponse', createComputerResponse_computer - A Computer object that represents the computer account.

$sel:httpStatus:CreateComputerResponse', createComputerResponse_httpStatus - The response's http status code.

Response Lenses

createComputerResponse_computer :: Lens' CreateComputerResponse (Maybe Computer)

A Computer object that represents the computer account.

createComputerResponse_httpStatus :: Lens' CreateComputerResponse Int

The response's http status code.