Amazonka.DirectoryService.CreateDirectory

Description

Creates a Simple AD directory. For more information, see Simple Active Directory in the Directory Service Admin Guide.

Before you call CreateDirectory, ensure that all of the required permissions have been explicitly granted through a policy. For details about what permissions are required to run the CreateDirectory operation, see Directory Service API Permissions: Actions, Resources, and Conditions Reference.

Creating a Request

data CreateDirectory

Contains the inputs for the CreateDirectory operation.

See: newCreateDirectory smart constructor.

Constructors

CreateDirectory'

Fields

shortName :: Maybe Text

The NetBIOS name of the directory, such as CORP.

vpcSettings :: Maybe DirectoryVpcSettings

A DirectoryVpcSettings object that contains additional information for the operation.

description :: Maybe Text

A description for the directory.

tags :: Maybe [Tag]

The tags to be assigned to the Simple AD directory.

name :: Text

The fully qualified name for the directory, such as corp.example.com.

password :: Sensitive Text

The password for the directory administrator. The directory creation process creates a directory administrator account with the user name Administrator and this password.

If you need to change the password for the administrator account, you can use the ResetUserPassword API call.

The regex pattern for this string is made up of the following conditions:

Length (?=^.{8,64}$) – Must be between 8 and 64 characters

AND any 3 of the following password complexity rules required by Active Directory:

Numbers and upper case and lowercase (?=.*\d)(?=.*[A-Z])(?=.*[a-z])

Numbers and special characters and lower case (?=.*\d)(?=.*[^A-Za-z0-9\s])(?=.*[a-z])

Special characters and upper case and lower case (?=.*[^A-Za-z0-9\s])(?=.*[A-Z])(?=.*[a-z])

Numbers and upper case and special characters (?=.*\d)(?=.*[A-Z])(?=.*[^A-Za-z0-9\s])

For additional information about how Active Directory passwords are enforced, see Password must meet complexity requirements on the Microsoft website.

size :: DirectorySize

The size of the directory.

newCreateDirectory

Arguments

:: Text
$sel:name:CreateDirectory'

-> Text
$sel:password:CreateDirectory'

-> DirectorySize
$sel:size:CreateDirectory'

-> CreateDirectory

Create a value of CreateDirectory with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:shortName:CreateDirectory', createDirectory_shortName - The NetBIOS name of the directory, such as CORP.

$sel:vpcSettings:CreateDirectory', createDirectory_vpcSettings - A DirectoryVpcSettings object that contains additional information for the operation.

$sel:description:CreateDirectory', createDirectory_description - A description for the directory.

$sel:tags:CreateDirectory', createDirectory_tags - The tags to be assigned to the Simple AD directory.

$sel:name:CreateDirectory', createDirectory_name - The fully qualified name for the directory, such as corp.example.com.

$sel:password:CreateDirectory', createDirectory_password - The password for the directory administrator. The directory creation process creates a directory administrator account with the user name Administrator and this password.

If you need to change the password for the administrator account, you can use the ResetUserPassword API call.

The regex pattern for this string is made up of the following conditions:

Length (?=^.{8,64}$) – Must be between 8 and 64 characters

AND any 3 of the following password complexity rules required by Active Directory:

Numbers and upper case and lowercase (?=.*\d)(?=.*[A-Z])(?=.*[a-z])

Numbers and special characters and lower case (?=.*\d)(?=.*[^A-Za-z0-9\s])(?=.*[a-z])

Special characters and upper case and lower case (?=.*[^A-Za-z0-9\s])(?=.*[A-Z])(?=.*[a-z])

Numbers and upper case and special characters (?=.*\d)(?=.*[A-Z])(?=.*[^A-Za-z0-9\s])

For additional information about how Active Directory passwords are enforced, see Password must meet complexity requirements on the Microsoft website.

$sel:size:CreateDirectory', createDirectory_size - The size of the directory.

Request Lenses

createDirectory_shortName :: Lens' CreateDirectory (Maybe Text)

The NetBIOS name of the directory, such as CORP.

createDirectory_vpcSettings :: Lens' CreateDirectory (Maybe DirectoryVpcSettings)

A DirectoryVpcSettings object that contains additional information for the operation.

createDirectory_description :: Lens' CreateDirectory (Maybe Text)

A description for the directory.

createDirectory_tags :: Lens' CreateDirectory (Maybe [Tag])

The tags to be assigned to the Simple AD directory.

createDirectory_name :: Lens' CreateDirectory Text

The fully qualified name for the directory, such as corp.example.com.

createDirectory_password :: Lens' CreateDirectory Text

The password for the directory administrator. The directory creation process creates a directory administrator account with the user name Administrator and this password.

If you need to change the password for the administrator account, you can use the ResetUserPassword API call.

The regex pattern for this string is made up of the following conditions:

Length (?=^.{8,64}$) – Must be between 8 and 64 characters

AND any 3 of the following password complexity rules required by Active Directory:

Numbers and upper case and lowercase (?=.*\d)(?=.*[A-Z])(?=.*[a-z])

Numbers and special characters and lower case (?=.*\d)(?=.*[^A-Za-z0-9\s])(?=.*[a-z])

Special characters and upper case and lower case (?=.*[^A-Za-z0-9\s])(?=.*[A-Z])(?=.*[a-z])

Numbers and upper case and special characters (?=.*\d)(?=.*[A-Z])(?=.*[^A-Za-z0-9\s])

For additional information about how Active Directory passwords are enforced, see Password must meet complexity requirements on the Microsoft website.

createDirectory_size :: Lens' CreateDirectory DirectorySize

The size of the directory.

Destructuring the Response

data CreateDirectoryResponse

Contains the results of the CreateDirectory operation.

See: newCreateDirectoryResponse smart constructor.

Constructors

CreateDirectoryResponse'

Fields

directoryId :: Maybe Text

The identifier of the directory that was created.

httpStatus :: Int

The response's http status code.

newCreateDirectoryResponse

Arguments

:: Int
$sel:httpStatus:CreateDirectoryResponse'

-> CreateDirectoryResponse

Create a value of CreateDirectoryResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:directoryId:CreateDirectoryResponse', createDirectoryResponse_directoryId - The identifier of the directory that was created.

$sel:httpStatus:CreateDirectoryResponse', createDirectoryResponse_httpStatus - The response's http status code.

Response Lenses

createDirectoryResponse_directoryId :: Lens' CreateDirectoryResponse (Maybe Text)

The identifier of the directory that was created.

createDirectoryResponse_httpStatus :: Lens' CreateDirectoryResponse Int

The response's http status code.