Amazonka.DirectoryService.CreateMicrosoftAD

Description

Creates a Microsoft AD directory in the Amazon Web Services Cloud. For more information, see Managed Microsoft AD in the Directory Service Admin Guide.

Before you call CreateMicrosoftAD, ensure that all of the required permissions have been explicitly granted through a policy. For details about what permissions are required to run the CreateMicrosoftAD operation, see Directory Service API Permissions: Actions, Resources, and Conditions Reference.

Creating a Request

data CreateMicrosoftAD

Creates an Managed Microsoft AD directory.

See: newCreateMicrosoftAD smart constructor.

Constructors

CreateMicrosoftAD'

Fields

edition :: Maybe DirectoryEdition

Managed Microsoft AD is available in two editions: Standard and Enterprise. Enterprise is the default.

shortName :: Maybe Text

The NetBIOS name for your domain, such as CORP. If you don't specify a NetBIOS name, it will default to the first part of your directory DNS. For example, CORP for the directory DNS corp.example.com.

description :: Maybe Text

A description for the directory. This label will appear on the Amazon Web Services console Directory Details page after the directory is created.

tags :: Maybe [Tag]

The tags to be assigned to the Managed Microsoft AD directory.

name :: Text

The fully qualified domain name for the Managed Microsoft AD directory, such as corp.example.com. This name will resolve inside your VPC only. It does not need to be publicly resolvable.

password :: Sensitive Text

The password for the default administrative user named Admin.

If you need to change the password for the administrator account, you can use the ResetUserPassword API call.

vpcSettings :: DirectoryVpcSettings

Contains VPC information for the CreateDirectory or CreateMicrosoftAD operation.

newCreateMicrosoftAD

Arguments

:: Text
$sel:name:CreateMicrosoftAD'

-> Text
$sel:password:CreateMicrosoftAD'

-> DirectoryVpcSettings
$sel:vpcSettings:CreateMicrosoftAD'

-> CreateMicrosoftAD

Create a value of CreateMicrosoftAD with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:edition:CreateMicrosoftAD', createMicrosoftAD_edition - Managed Microsoft AD is available in two editions: Standard and Enterprise. Enterprise is the default.

$sel:shortName:CreateMicrosoftAD', createMicrosoftAD_shortName - The NetBIOS name for your domain, such as CORP. If you don't specify a NetBIOS name, it will default to the first part of your directory DNS. For example, CORP for the directory DNS corp.example.com.

$sel:description:CreateMicrosoftAD', createMicrosoftAD_description - A description for the directory. This label will appear on the Amazon Web Services console Directory Details page after the directory is created.

$sel:tags:CreateMicrosoftAD', createMicrosoftAD_tags - The tags to be assigned to the Managed Microsoft AD directory.

$sel:name:CreateMicrosoftAD', createMicrosoftAD_name - The fully qualified domain name for the Managed Microsoft AD directory, such as corp.example.com. This name will resolve inside your VPC only. It does not need to be publicly resolvable.

$sel:password:CreateMicrosoftAD', createMicrosoftAD_password - The password for the default administrative user named Admin.

If you need to change the password for the administrator account, you can use the ResetUserPassword API call.

$sel:vpcSettings:CreateMicrosoftAD', createMicrosoftAD_vpcSettings - Contains VPC information for the CreateDirectory or CreateMicrosoftAD operation.

Request Lenses

createMicrosoftAD_edition :: Lens' CreateMicrosoftAD (Maybe DirectoryEdition)

Managed Microsoft AD is available in two editions: Standard and Enterprise. Enterprise is the default.

createMicrosoftAD_shortName :: Lens' CreateMicrosoftAD (Maybe Text)

The NetBIOS name for your domain, such as CORP. If you don't specify a NetBIOS name, it will default to the first part of your directory DNS. For example, CORP for the directory DNS corp.example.com.

createMicrosoftAD_description :: Lens' CreateMicrosoftAD (Maybe Text)

A description for the directory. This label will appear on the Amazon Web Services console Directory Details page after the directory is created.

createMicrosoftAD_tags :: Lens' CreateMicrosoftAD (Maybe [Tag])

The tags to be assigned to the Managed Microsoft AD directory.

createMicrosoftAD_name :: Lens' CreateMicrosoftAD Text

The fully qualified domain name for the Managed Microsoft AD directory, such as corp.example.com. This name will resolve inside your VPC only. It does not need to be publicly resolvable.

createMicrosoftAD_password :: Lens' CreateMicrosoftAD Text

The password for the default administrative user named Admin.

If you need to change the password for the administrator account, you can use the ResetUserPassword API call.

createMicrosoftAD_vpcSettings :: Lens' CreateMicrosoftAD DirectoryVpcSettings

Contains VPC information for the CreateDirectory or CreateMicrosoftAD operation.

Destructuring the Response

data CreateMicrosoftADResponse

Result of a CreateMicrosoftAD request.

See: newCreateMicrosoftADResponse smart constructor.

Constructors

CreateMicrosoftADResponse'

Fields

directoryId :: Maybe Text

The identifier of the directory that was created.

httpStatus :: Int

The response's http status code.

newCreateMicrosoftADResponse

Arguments

:: Int
$sel:httpStatus:CreateMicrosoftADResponse'

-> CreateMicrosoftADResponse

Create a value of CreateMicrosoftADResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:directoryId:CreateMicrosoftADResponse', createMicrosoftADResponse_directoryId - The identifier of the directory that was created.

$sel:httpStatus:CreateMicrosoftADResponse', createMicrosoftADResponse_httpStatus - The response's http status code.

Response Lenses

createMicrosoftADResponse_directoryId :: Lens' CreateMicrosoftADResponse (Maybe Text)

The identifier of the directory that was created.

createMicrosoftADResponse_httpStatus :: Lens' CreateMicrosoftADResponse Int

The response's http status code.