Amazonka.DirectoryService.DescribeConditionalForwarders

Description

Obtains information about the conditional forwarders for this account.

If no input parameters are provided for RemoteDomainNames, this request describes all conditional forwarders for the specified directory ID.

Creating a Request

data DescribeConditionalForwarders

Describes a conditional forwarder.

See: newDescribeConditionalForwarders smart constructor.

Constructors

DescribeConditionalForwarders'

Fields

remoteDomainNames :: Maybe [Text]

The fully qualified domain names (FQDN) of the remote domains for which to get the list of associated conditional forwarders. If this member is null, all conditional forwarders are returned.

directoryId :: Text

The directory ID for which to get the list of associated conditional forwarders.

newDescribeConditionalForwarders

Arguments

:: Text
$sel:directoryId:DescribeConditionalForwarders'

-> DescribeConditionalForwarders

Create a value of DescribeConditionalForwarders with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:remoteDomainNames:DescribeConditionalForwarders', describeConditionalForwarders_remoteDomainNames - The fully qualified domain names (FQDN) of the remote domains for which to get the list of associated conditional forwarders. If this member is null, all conditional forwarders are returned.

$sel:directoryId:DescribeConditionalForwarders', describeConditionalForwarders_directoryId - The directory ID for which to get the list of associated conditional forwarders.

Request Lenses

describeConditionalForwarders_remoteDomainNames :: Lens' DescribeConditionalForwarders (Maybe [Text])

The fully qualified domain names (FQDN) of the remote domains for which to get the list of associated conditional forwarders. If this member is null, all conditional forwarders are returned.

describeConditionalForwarders_directoryId :: Lens' DescribeConditionalForwarders Text

The directory ID for which to get the list of associated conditional forwarders.

Destructuring the Response

data DescribeConditionalForwardersResponse

The result of a DescribeConditionalForwarder request.

See: newDescribeConditionalForwardersResponse smart constructor.

Constructors

DescribeConditionalForwardersResponse'

Fields

conditionalForwarders :: Maybe [ConditionalForwarder]

The list of conditional forwarders that have been created.

httpStatus :: Int

The response's http status code.

newDescribeConditionalForwardersResponse

Arguments

:: Int
$sel:httpStatus:DescribeConditionalForwardersResponse'

-> DescribeConditionalForwardersResponse

Create a value of DescribeConditionalForwardersResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:conditionalForwarders:DescribeConditionalForwardersResponse', describeConditionalForwardersResponse_conditionalForwarders - The list of conditional forwarders that have been created.

$sel:httpStatus:DescribeConditionalForwardersResponse', describeConditionalForwardersResponse_httpStatus - The response's http status code.

Response Lenses

describeConditionalForwardersResponse_conditionalForwarders :: Lens' DescribeConditionalForwardersResponse (Maybe [ConditionalForwarder])

The list of conditional forwarders that have been created.

describeConditionalForwardersResponse_httpStatus :: Lens' DescribeConditionalForwardersResponse Int

The response's http status code.