Amazonka.DirectoryService.Types.Certificate

Description

Documentation

data Certificate

Information about the certificate.

See: newCertificate smart constructor.

Constructors

Certificate'

Fields

clientCertAuthSettings :: Maybe ClientCertAuthSettings

A ClientCertAuthSettings object that contains client certificate authentication settings.

state :: Maybe CertificateState

The state of the certificate.

commonName :: Maybe Text

The common name for the certificate.

certificateId :: Maybe Text

The identifier of the certificate.

expiryDateTime :: Maybe POSIX

The date and time when the certificate will expire.

registeredDateTime :: Maybe POSIX

The date and time that the certificate was registered.

type' :: Maybe CertificateType

The function that the registered certificate performs. Valid values include ClientLDAPS or ClientCertAuth. The default value is ClientLDAPS.

stateReason :: Maybe Text

Describes a state change for the certificate.

newCertificate :: Certificate

Create a value of Certificate with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:clientCertAuthSettings:Certificate', certificate_clientCertAuthSettings - A ClientCertAuthSettings object that contains client certificate authentication settings.

$sel:state:Certificate', certificate_state - The state of the certificate.

$sel:commonName:Certificate', certificate_commonName - The common name for the certificate.

$sel:certificateId:Certificate', certificate_certificateId - The identifier of the certificate.

$sel:expiryDateTime:Certificate', certificate_expiryDateTime - The date and time when the certificate will expire.

$sel:registeredDateTime:Certificate', certificate_registeredDateTime - The date and time that the certificate was registered.

$sel:type':Certificate', certificate_type - The function that the registered certificate performs. Valid values include ClientLDAPS or ClientCertAuth. The default value is ClientLDAPS.

$sel:stateReason:Certificate', certificate_stateReason - Describes a state change for the certificate.

certificate_clientCertAuthSettings :: Lens' Certificate (Maybe ClientCertAuthSettings)

A ClientCertAuthSettings object that contains client certificate authentication settings.

certificate_state :: Lens' Certificate (Maybe CertificateState)

The state of the certificate.

certificate_commonName :: Lens' Certificate (Maybe Text)

The common name for the certificate.

certificate_certificateId :: Lens' Certificate (Maybe Text)

The identifier of the certificate.

certificate_expiryDateTime :: Lens' Certificate (Maybe UTCTime)

The date and time when the certificate will expire.

certificate_registeredDateTime :: Lens' Certificate (Maybe UTCTime)

The date and time that the certificate was registered.

certificate_type :: Lens' Certificate (Maybe CertificateType)

The function that the registered certificate performs. Valid values include ClientLDAPS or ClientCertAuth. The default value is ClientLDAPS.

certificate_stateReason :: Lens' Certificate (Maybe Text)

Describes a state change for the certificate.