Amazonka.DirectoryService.Types.DomainController

Description

Documentation

data DomainController

Contains information about the domain controllers for a specified directory.

See: newDomainController smart constructor.

Constructors

DomainController'

Fields

status :: Maybe DomainControllerStatus

The status of the domain controller.

directoryId :: Maybe Text

Identifier of the directory where the domain controller resides.

vpcId :: Maybe Text

The identifier of the VPC that contains the domain controller.

launchTime :: Maybe POSIX

Specifies when the domain controller was created.

subnetId :: Maybe Text

Identifier of the subnet in the VPC that contains the domain controller.

availabilityZone :: Maybe Text

The Availability Zone where the domain controller is located.

statusLastUpdatedDateTime :: Maybe POSIX

The date and time that the status was last updated.

statusReason :: Maybe Text

A description of the domain controller state.

dnsIpAddr :: Maybe Text

The IP address of the domain controller.

domainControllerId :: Maybe Text

Identifies a specific domain controller in the directory.

newDomainController :: DomainController

Create a value of DomainController with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:status:DomainController', domainController_status - The status of the domain controller.

$sel:directoryId:DomainController', domainController_directoryId - Identifier of the directory where the domain controller resides.

$sel:vpcId:DomainController', domainController_vpcId - The identifier of the VPC that contains the domain controller.

$sel:launchTime:DomainController', domainController_launchTime - Specifies when the domain controller was created.

$sel:subnetId:DomainController', domainController_subnetId - Identifier of the subnet in the VPC that contains the domain controller.

$sel:availabilityZone:DomainController', domainController_availabilityZone - The Availability Zone where the domain controller is located.

$sel:statusLastUpdatedDateTime:DomainController', domainController_statusLastUpdatedDateTime - The date and time that the status was last updated.

$sel:statusReason:DomainController', domainController_statusReason - A description of the domain controller state.

$sel:dnsIpAddr:DomainController', domainController_dnsIpAddr - The IP address of the domain controller.

$sel:domainControllerId:DomainController', domainController_domainControllerId - Identifies a specific domain controller in the directory.

domainController_status :: Lens' DomainController (Maybe DomainControllerStatus)

The status of the domain controller.

domainController_directoryId :: Lens' DomainController (Maybe Text)

Identifier of the directory where the domain controller resides.

domainController_vpcId :: Lens' DomainController (Maybe Text)

The identifier of the VPC that contains the domain controller.

domainController_launchTime :: Lens' DomainController (Maybe UTCTime)

Specifies when the domain controller was created.

domainController_subnetId :: Lens' DomainController (Maybe Text)

Identifier of the subnet in the VPC that contains the domain controller.

domainController_availabilityZone :: Lens' DomainController (Maybe Text)

The Availability Zone where the domain controller is located.

domainController_statusLastUpdatedDateTime :: Lens' DomainController (Maybe UTCTime)

The date and time that the status was last updated.

domainController_statusReason :: Lens' DomainController (Maybe Text)

A description of the domain controller state.

domainController_dnsIpAddr :: Lens' DomainController (Maybe Text)

The IP address of the domain controller.

domainController_domainControllerId :: Lens' DomainController (Maybe Text)

Identifies a specific domain controller in the directory.