Amazonka.DirectoryService.Types.RegionDescription

Description

Documentation

data RegionDescription

The replicated Region information for a directory.

See: newRegionDescription smart constructor.

Constructors

RegionDescription'

Fields

status :: Maybe DirectoryStage

The status of the replication process for the specified Region.

directoryId :: Maybe Text

The identifier of the directory.

regionName :: Maybe Text

The name of the Region. For example, us-east-1.

desiredNumberOfDomainControllers :: Maybe Natural

The desired number of domain controllers in the specified Region for the specified directory.

regionType :: Maybe RegionType

Specifies whether the Region is the primary Region or an additional Region.

launchTime :: Maybe POSIX

Specifies when the Region replication began.

lastUpdatedDateTime :: Maybe POSIX

The date and time that the Region description was last updated.

statusLastUpdatedDateTime :: Maybe POSIX

The date and time that the Region status was last updated.

vpcSettings :: Maybe DirectoryVpcSettings

newRegionDescription :: RegionDescription

Create a value of RegionDescription with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:status:RegionDescription', regionDescription_status - The status of the replication process for the specified Region.

$sel:directoryId:RegionDescription', regionDescription_directoryId - The identifier of the directory.

$sel:regionName:RegionDescription', regionDescription_regionName - The name of the Region. For example, us-east-1.

$sel:desiredNumberOfDomainControllers:RegionDescription', regionDescription_desiredNumberOfDomainControllers - The desired number of domain controllers in the specified Region for the specified directory.

$sel:regionType:RegionDescription', regionDescription_regionType - Specifies whether the Region is the primary Region or an additional Region.

$sel:launchTime:RegionDescription', regionDescription_launchTime - Specifies when the Region replication began.

$sel:lastUpdatedDateTime:RegionDescription', regionDescription_lastUpdatedDateTime - The date and time that the Region description was last updated.

$sel:statusLastUpdatedDateTime:RegionDescription', regionDescription_statusLastUpdatedDateTime - The date and time that the Region status was last updated.

$sel:vpcSettings:RegionDescription', regionDescription_vpcSettings - Undocumented member.

regionDescription_status :: Lens' RegionDescription (Maybe DirectoryStage)

The status of the replication process for the specified Region.

regionDescription_directoryId :: Lens' RegionDescription (Maybe Text)

The identifier of the directory.

regionDescription_regionName :: Lens' RegionDescription (Maybe Text)

The name of the Region. For example, us-east-1.

regionDescription_desiredNumberOfDomainControllers :: Lens' RegionDescription (Maybe Natural)

The desired number of domain controllers in the specified Region for the specified directory.

regionDescription_regionType :: Lens' RegionDescription (Maybe RegionType)

Specifies whether the Region is the primary Region or an additional Region.

regionDescription_launchTime :: Lens' RegionDescription (Maybe UTCTime)

Specifies when the Region replication began.

regionDescription_lastUpdatedDateTime :: Lens' RegionDescription (Maybe UTCTime)

The date and time that the Region description was last updated.

regionDescription_statusLastUpdatedDateTime :: Lens' RegionDescription (Maybe UTCTime)

The date and time that the Region status was last updated.

regionDescription_vpcSettings :: Lens' RegionDescription (Maybe DirectoryVpcSettings)

Undocumented member.