Amazonka.DirectoryService.Types.SharedDirectory

Description

Documentation

data SharedDirectory

Details about the shared directory in the directory owner account for which the share request in the directory consumer account has been accepted.

See: newSharedDirectory smart constructor.

Constructors

SharedDirectory'

Fields

sharedAccountId :: Maybe Text

Identifier of the directory consumer account that has access to the shared directory (OwnerDirectoryId) in the directory owner account.

ownerAccountId :: Maybe Text

Identifier of the directory owner account, which contains the directory that has been shared to the consumer account.

lastUpdatedDateTime :: Maybe POSIX

The date and time that the shared directory was last updated.

shareStatus :: Maybe ShareStatus

Current directory status of the shared Managed Microsoft AD directory.

shareMethod :: Maybe ShareMethod

The method used when sharing a directory to determine whether the directory should be shared within your Amazon Web Services organization (ORGANIZATIONS) or with any Amazon Web Services account by sending a shared directory request (HANDSHAKE).

ownerDirectoryId :: Maybe Text

Identifier of the directory in the directory owner account.

sharedDirectoryId :: Maybe Text

Identifier of the shared directory in the directory consumer account. This identifier is different for each directory owner account.

shareNotes :: Maybe (Sensitive Text)

A directory share request that is sent by the directory owner to the directory consumer. The request includes a typed message to help the directory consumer administrator determine whether to approve or reject the share invitation.

createdDateTime :: Maybe POSIX

The date and time that the shared directory was created.

newSharedDirectory :: SharedDirectory

Create a value of SharedDirectory with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:sharedAccountId:SharedDirectory', sharedDirectory_sharedAccountId - Identifier of the directory consumer account that has access to the shared directory (OwnerDirectoryId) in the directory owner account.

$sel:ownerAccountId:SharedDirectory', sharedDirectory_ownerAccountId - Identifier of the directory owner account, which contains the directory that has been shared to the consumer account.

$sel:lastUpdatedDateTime:SharedDirectory', sharedDirectory_lastUpdatedDateTime - The date and time that the shared directory was last updated.

$sel:shareStatus:SharedDirectory', sharedDirectory_shareStatus - Current directory status of the shared Managed Microsoft AD directory.

$sel:shareMethod:SharedDirectory', sharedDirectory_shareMethod - The method used when sharing a directory to determine whether the directory should be shared within your Amazon Web Services organization (ORGANIZATIONS) or with any Amazon Web Services account by sending a shared directory request (HANDSHAKE).

$sel:ownerDirectoryId:SharedDirectory', sharedDirectory_ownerDirectoryId - Identifier of the directory in the directory owner account.

$sel:sharedDirectoryId:SharedDirectory', sharedDirectory_sharedDirectoryId - Identifier of the shared directory in the directory consumer account. This identifier is different for each directory owner account.

$sel:shareNotes:SharedDirectory', sharedDirectory_shareNotes - A directory share request that is sent by the directory owner to the directory consumer. The request includes a typed message to help the directory consumer administrator determine whether to approve or reject the share invitation.

$sel:createdDateTime:SharedDirectory', sharedDirectory_createdDateTime - The date and time that the shared directory was created.

sharedDirectory_sharedAccountId :: Lens' SharedDirectory (Maybe Text)

Identifier of the directory consumer account that has access to the shared directory (OwnerDirectoryId) in the directory owner account.

sharedDirectory_ownerAccountId :: Lens' SharedDirectory (Maybe Text)

Identifier of the directory owner account, which contains the directory that has been shared to the consumer account.

sharedDirectory_lastUpdatedDateTime :: Lens' SharedDirectory (Maybe UTCTime)

The date and time that the shared directory was last updated.

sharedDirectory_shareStatus :: Lens' SharedDirectory (Maybe ShareStatus)

Current directory status of the shared Managed Microsoft AD directory.

sharedDirectory_shareMethod :: Lens' SharedDirectory (Maybe ShareMethod)

The method used when sharing a directory to determine whether the directory should be shared within your Amazon Web Services organization (ORGANIZATIONS) or with any Amazon Web Services account by sending a shared directory request (HANDSHAKE).

sharedDirectory_ownerDirectoryId :: Lens' SharedDirectory (Maybe Text)

Identifier of the directory in the directory owner account.

sharedDirectory_sharedDirectoryId :: Lens' SharedDirectory (Maybe Text)

Identifier of the shared directory in the directory consumer account. This identifier is different for each directory owner account.

sharedDirectory_shareNotes :: Lens' SharedDirectory (Maybe Text)

A directory share request that is sent by the directory owner to the directory consumer. The request includes a typed message to help the directory consumer administrator determine whether to approve or reject the share invitation.

sharedDirectory_createdDateTime :: Lens' SharedDirectory (Maybe UTCTime)

The date and time that the shared directory was created.