Amazonka.DirectoryService.VerifyTrust

Description

Directory Service for Microsoft Active Directory allows you to configure and verify trust relationships.

This action verifies a trust relationship between your Managed Microsoft AD directory and an external domain.

Creating a Request

data VerifyTrust

Initiates the verification of an existing trust relationship between an Managed Microsoft AD directory and an external domain.

See: newVerifyTrust smart constructor.

Constructors

VerifyTrust'

Fields

trustId :: Text

The unique Trust ID of the trust relationship to verify.

newVerifyTrust

Arguments

:: Text
$sel:trustId:VerifyTrust'

-> VerifyTrust

Create a value of VerifyTrust with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:trustId:VerifyTrust', verifyTrust_trustId - The unique Trust ID of the trust relationship to verify.

Request Lenses

verifyTrust_trustId :: Lens' VerifyTrust Text

The unique Trust ID of the trust relationship to verify.

Destructuring the Response

data VerifyTrustResponse

Result of a VerifyTrust request.

See: newVerifyTrustResponse smart constructor.

Constructors

VerifyTrustResponse'

Fields

trustId :: Maybe Text

The unique Trust ID of the trust relationship that was verified.

httpStatus :: Int

The response's http status code.

newVerifyTrustResponse

Arguments

:: Int
$sel:httpStatus:VerifyTrustResponse'

-> VerifyTrustResponse

Create a value of VerifyTrustResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:trustId:VerifyTrust', verifyTrustResponse_trustId - The unique Trust ID of the trust relationship that was verified.

$sel:httpStatus:VerifyTrustResponse', verifyTrustResponse_httpStatus - The response's http status code.

Response Lenses

verifyTrustResponse_trustId :: Lens' VerifyTrustResponse (Maybe Text)

The unique Trust ID of the trust relationship that was verified.

verifyTrustResponse_httpStatus :: Lens' VerifyTrustResponse Int

The response's http status code.