Amazonka.Discovery.DescribeAgents

Description

Lists agents or connectors as specified by ID or other filters. All agents/connectors associated with your user account can be listed if you call DescribeAgents as is without passing any parameters.

This operation returns paginated results.

Creating a Request

data DescribeAgents

See: newDescribeAgents smart constructor.

Constructors

DescribeAgents'

Fields

agentIds :: Maybe [Text]

The agent or the Connector IDs for which you want information. If you specify no IDs, the system returns information about all agents/Connectors associated with your AWS user account.

filters :: Maybe [Filter]

You can filter the request using various logical operators and a key-value format. For example:

{"key": "collectionStatus", "value": "STARTED"}

nextToken :: Maybe Text

Token to retrieve the next set of results. For example, if you previously specified 100 IDs for DescribeAgentsRequest$agentIds but set DescribeAgentsRequest$maxResults to 10, you received a set of 10 results along with a token. Use that token in this query to get the next set of 10.

maxResults :: Maybe Int

The total number of agents/Connectors to return in a single page of output. The maximum value is 100.

newDescribeAgents :: DescribeAgents

Create a value of DescribeAgents with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:agentIds:DescribeAgents', describeAgents_agentIds - The agent or the Connector IDs for which you want information. If you specify no IDs, the system returns information about all agents/Connectors associated with your AWS user account.

$sel:filters:DescribeAgents', describeAgents_filters - You can filter the request using various logical operators and a key-value format. For example:

{"key": "collectionStatus", "value": "STARTED"}

$sel:nextToken:DescribeAgents', describeAgents_nextToken - Token to retrieve the next set of results. For example, if you previously specified 100 IDs for DescribeAgentsRequest$agentIds but set DescribeAgentsRequest$maxResults to 10, you received a set of 10 results along with a token. Use that token in this query to get the next set of 10.

$sel:maxResults:DescribeAgents', describeAgents_maxResults - The total number of agents/Connectors to return in a single page of output. The maximum value is 100.

Request Lenses

describeAgents_agentIds :: Lens' DescribeAgents (Maybe [Text])

The agent or the Connector IDs for which you want information. If you specify no IDs, the system returns information about all agents/Connectors associated with your AWS user account.

describeAgents_filters :: Lens' DescribeAgents (Maybe [Filter])

You can filter the request using various logical operators and a key-value format. For example:

{"key": "collectionStatus", "value": "STARTED"}

describeAgents_nextToken :: Lens' DescribeAgents (Maybe Text)

Token to retrieve the next set of results. For example, if you previously specified 100 IDs for DescribeAgentsRequest$agentIds but set DescribeAgentsRequest$maxResults to 10, you received a set of 10 results along with a token. Use that token in this query to get the next set of 10.

describeAgents_maxResults :: Lens' DescribeAgents (Maybe Int)

The total number of agents/Connectors to return in a single page of output. The maximum value is 100.

Destructuring the Response

data DescribeAgentsResponse

See: newDescribeAgentsResponse smart constructor.

Constructors

DescribeAgentsResponse'

Fields

agentsInfo :: Maybe [AgentInfo]

Lists agents or the Connector by ID or lists all agents/Connectors associated with your user account if you did not specify an agent/Connector ID. The output includes agent/Connector IDs, IP addresses, media access control (MAC) addresses, agent/Connector health, host name where the agent/Connector resides, and the version number of each agent/Connector.

nextToken :: Maybe Text

Token to retrieve the next set of results. For example, if you specified 100 IDs for DescribeAgentsRequest$agentIds but set DescribeAgentsRequest$maxResults to 10, you received a set of 10 results along with this token. Use this token in the next query to retrieve the next set of 10.

httpStatus :: Int

The response's http status code.

newDescribeAgentsResponse

Arguments

:: Int
$sel:httpStatus:DescribeAgentsResponse'

-> DescribeAgentsResponse

Create a value of DescribeAgentsResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:agentsInfo:DescribeAgentsResponse', describeAgentsResponse_agentsInfo - Lists agents or the Connector by ID or lists all agents/Connectors associated with your user account if you did not specify an agent/Connector ID. The output includes agent/Connector IDs, IP addresses, media access control (MAC) addresses, agent/Connector health, host name where the agent/Connector resides, and the version number of each agent/Connector.

$sel:nextToken:DescribeAgents', describeAgentsResponse_nextToken - Token to retrieve the next set of results. For example, if you specified 100 IDs for DescribeAgentsRequest$agentIds but set DescribeAgentsRequest$maxResults to 10, you received a set of 10 results along with this token. Use this token in the next query to retrieve the next set of 10.

$sel:httpStatus:DescribeAgentsResponse', describeAgentsResponse_httpStatus - The response's http status code.

Response Lenses

describeAgentsResponse_agentsInfo :: Lens' DescribeAgentsResponse (Maybe [AgentInfo])

Lists agents or the Connector by ID or lists all agents/Connectors associated with your user account if you did not specify an agent/Connector ID. The output includes agent/Connector IDs, IP addresses, media access control (MAC) addresses, agent/Connector health, host name where the agent/Connector resides, and the version number of each agent/Connector.

describeAgentsResponse_nextToken :: Lens' DescribeAgentsResponse (Maybe Text)

Token to retrieve the next set of results. For example, if you specified 100 IDs for DescribeAgentsRequest$agentIds but set DescribeAgentsRequest$maxResults to 10, you received a set of 10 results along with this token. Use this token in the next query to retrieve the next set of 10.

describeAgentsResponse_httpStatus :: Lens' DescribeAgentsResponse Int

The response's http status code.