Amazonka.Discovery.DescribeTags

Description

Retrieves a list of configuration items that have tags as specified by the key-value pairs, name and value, passed to the optional parameter filters.

There are three valid tag filter names:

tagKey

tagValue

configurationId

Also, all configuration items associated with your user account that have tags can be listed if you call DescribeTags as is without passing any parameters.

This operation returns paginated results.

Creating a Request

data DescribeTags

See: newDescribeTags smart constructor.

Constructors

DescribeTags'

Fields

filters :: Maybe [TagFilter]

You can filter the list using a key-value format. You can separate these items by using logical operators. Allowed filters include tagKey, tagValue, and configurationId.

nextToken :: Maybe Text

A token to start the list. Use this token to get the next set of results.

maxResults :: Maybe Int

The total number of items to return in a single page of output. The maximum value is 100.

newDescribeTags :: DescribeTags

Create a value of DescribeTags with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:filters:DescribeTags', describeTags_filters - You can filter the list using a key-value format. You can separate these items by using logical operators. Allowed filters include tagKey, tagValue, and configurationId.

$sel:nextToken:DescribeTags', describeTags_nextToken - A token to start the list. Use this token to get the next set of results.

$sel:maxResults:DescribeTags', describeTags_maxResults - The total number of items to return in a single page of output. The maximum value is 100.

Request Lenses

describeTags_filters :: Lens' DescribeTags (Maybe [TagFilter])

You can filter the list using a key-value format. You can separate these items by using logical operators. Allowed filters include tagKey, tagValue, and configurationId.

describeTags_nextToken :: Lens' DescribeTags (Maybe Text)

A token to start the list. Use this token to get the next set of results.

describeTags_maxResults :: Lens' DescribeTags (Maybe Int)

The total number of items to return in a single page of output. The maximum value is 100.

Destructuring the Response

data DescribeTagsResponse

See: newDescribeTagsResponse smart constructor.

Constructors

DescribeTagsResponse'

Fields

nextToken :: Maybe Text

The call returns a token. Use this token to get the next set of results.

tags :: Maybe [ConfigurationTag]

Depending on the input, this is a list of configuration items tagged with a specific tag, or a list of tags for a specific configuration item.

httpStatus :: Int

The response's http status code.

newDescribeTagsResponse

Arguments

:: Int
$sel:httpStatus:DescribeTagsResponse'

-> DescribeTagsResponse

Create a value of DescribeTagsResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:nextToken:DescribeTags', describeTagsResponse_nextToken - The call returns a token. Use this token to get the next set of results.

$sel:tags:DescribeTagsResponse', describeTagsResponse_tags - Depending on the input, this is a list of configuration items tagged with a specific tag, or a list of tags for a specific configuration item.

$sel:httpStatus:DescribeTagsResponse', describeTagsResponse_httpStatus - The response's http status code.

Response Lenses

describeTagsResponse_nextToken :: Lens' DescribeTagsResponse (Maybe Text)

The call returns a token. Use this token to get the next set of results.

describeTagsResponse_tags :: Lens' DescribeTagsResponse (Maybe [ConfigurationTag])

Depending on the input, this is a list of configuration items tagged with a specific tag, or a list of tags for a specific configuration item.

describeTagsResponse_httpStatus :: Lens' DescribeTagsResponse Int

The response's http status code.