Amazonka.Discovery.ListConfigurations

Description

Retrieves a list of configuration items as specified by the value passed to the required parameter configurationType. Optional filtering may be applied to refine search results.

This operation returns paginated results.

Creating a Request

data ListConfigurations

See: newListConfigurations smart constructor.

Constructors

ListConfigurations'

Fields

orderBy :: Maybe [OrderByElement]

Certain filter criteria return output that can be sorted in ascending or descending order. For a list of output characteristics for each filter, see Using the ListConfigurations Action in the AWS Application Discovery Service User Guide.

filters :: Maybe [Filter]

You can filter the request using various logical operators and a key-value format. For example:

{"key": "serverType", "value": "webServer"}

For a complete list of filter options and guidance about using them with this action, see Using the ListConfigurations Action in the AWS Application Discovery Service User Guide.

nextToken :: Maybe Text

Token to retrieve the next set of results. For example, if a previous call to ListConfigurations returned 100 items, but you set ListConfigurationsRequest$maxResults to 10, you received a set of 10 results along with a token. Use that token in this query to get the next set of 10.

maxResults :: Maybe Int

The total number of items to return. The maximum value is 100.

configurationType :: ConfigurationItemType

A valid configuration identified by Application Discovery Service.

newListConfigurations

Arguments

:: ConfigurationItemType
$sel:configurationType:ListConfigurations'

-> ListConfigurations

Create a value of ListConfigurations with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:orderBy:ListConfigurations', listConfigurations_orderBy - Certain filter criteria return output that can be sorted in ascending or descending order. For a list of output characteristics for each filter, see Using the ListConfigurations Action in the AWS Application Discovery Service User Guide.

$sel:filters:ListConfigurations', listConfigurations_filters - You can filter the request using various logical operators and a key-value format. For example:

{"key": "serverType", "value": "webServer"}

For a complete list of filter options and guidance about using them with this action, see Using the ListConfigurations Action in the AWS Application Discovery Service User Guide.

$sel:nextToken:ListConfigurations', listConfigurations_nextToken - Token to retrieve the next set of results. For example, if a previous call to ListConfigurations returned 100 items, but you set ListConfigurationsRequest$maxResults to 10, you received a set of 10 results along with a token. Use that token in this query to get the next set of 10.

$sel:maxResults:ListConfigurations', listConfigurations_maxResults - The total number of items to return. The maximum value is 100.

$sel:configurationType:ListConfigurations', listConfigurations_configurationType - A valid configuration identified by Application Discovery Service.

Request Lenses

listConfigurations_orderBy :: Lens' ListConfigurations (Maybe [OrderByElement])

Certain filter criteria return output that can be sorted in ascending or descending order. For a list of output characteristics for each filter, see Using the ListConfigurations Action in the AWS Application Discovery Service User Guide.

listConfigurations_filters :: Lens' ListConfigurations (Maybe [Filter])

You can filter the request using various logical operators and a key-value format. For example:

{"key": "serverType", "value": "webServer"}

For a complete list of filter options and guidance about using them with this action, see Using the ListConfigurations Action in the AWS Application Discovery Service User Guide.

listConfigurations_nextToken :: Lens' ListConfigurations (Maybe Text)

Token to retrieve the next set of results. For example, if a previous call to ListConfigurations returned 100 items, but you set ListConfigurationsRequest$maxResults to 10, you received a set of 10 results along with a token. Use that token in this query to get the next set of 10.

listConfigurations_maxResults :: Lens' ListConfigurations (Maybe Int)

The total number of items to return. The maximum value is 100.

listConfigurations_configurationType :: Lens' ListConfigurations ConfigurationItemType

A valid configuration identified by Application Discovery Service.

Destructuring the Response

data ListConfigurationsResponse

See: newListConfigurationsResponse smart constructor.

Constructors

ListConfigurationsResponse'

Fields

configurations :: Maybe [HashMap Text Text]

Returns configuration details, including the configuration ID, attribute names, and attribute values.

nextToken :: Maybe Text

Token to retrieve the next set of results. For example, if your call to ListConfigurations returned 100 items, but you set ListConfigurationsRequest$maxResults to 10, you received a set of 10 results along with this token. Use this token in the next query to retrieve the next set of 10.

httpStatus :: Int

The response's http status code.

newListConfigurationsResponse

Arguments

:: Int
$sel:httpStatus:ListConfigurationsResponse'

-> ListConfigurationsResponse

Create a value of ListConfigurationsResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:configurations:ListConfigurationsResponse', listConfigurationsResponse_configurations - Returns configuration details, including the configuration ID, attribute names, and attribute values.

$sel:nextToken:ListConfigurations', listConfigurationsResponse_nextToken - Token to retrieve the next set of results. For example, if your call to ListConfigurations returned 100 items, but you set ListConfigurationsRequest$maxResults to 10, you received a set of 10 results along with this token. Use this token in the next query to retrieve the next set of 10.

$sel:httpStatus:ListConfigurationsResponse', listConfigurationsResponse_httpStatus - The response's http status code.

Response Lenses

listConfigurationsResponse_configurations :: Lens' ListConfigurationsResponse (Maybe [HashMap Text Text])

Returns configuration details, including the configuration ID, attribute names, and attribute values.

listConfigurationsResponse_nextToken :: Lens' ListConfigurationsResponse (Maybe Text)

Token to retrieve the next set of results. For example, if your call to ListConfigurations returned 100 items, but you set ListConfigurationsRequest$maxResults to 10, you received a set of 10 results along with this token. Use this token in the next query to retrieve the next set of 10.

listConfigurationsResponse_httpStatus :: Lens' ListConfigurationsResponse Int

The response's http status code.