Amazonka.Discovery.Types.BatchDeleteImportDataError

Description

Documentation

data BatchDeleteImportDataError

Error messages returned for each import task that you deleted as a response for this command.

See: newBatchDeleteImportDataError smart constructor.

Constructors

BatchDeleteImportDataError'

Fields

importTaskId :: Maybe Text

The unique import ID associated with the error that occurred.

errorCode :: Maybe BatchDeleteImportDataErrorCode

The type of error that occurred for a specific import task.

errorDescription :: Maybe Text

The description of the error that occurred for a specific import task.

newBatchDeleteImportDataError :: BatchDeleteImportDataError

Create a value of BatchDeleteImportDataError with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:importTaskId:BatchDeleteImportDataError', batchDeleteImportDataError_importTaskId - The unique import ID associated with the error that occurred.

$sel:errorCode:BatchDeleteImportDataError', batchDeleteImportDataError_errorCode - The type of error that occurred for a specific import task.

$sel:errorDescription:BatchDeleteImportDataError', batchDeleteImportDataError_errorDescription - The description of the error that occurred for a specific import task.

batchDeleteImportDataError_importTaskId :: Lens' BatchDeleteImportDataError (Maybe Text)

The unique import ID associated with the error that occurred.

batchDeleteImportDataError_errorCode :: Lens' BatchDeleteImportDataError (Maybe BatchDeleteImportDataErrorCode)

The type of error that occurred for a specific import task.

batchDeleteImportDataError_errorDescription :: Lens' BatchDeleteImportDataError (Maybe Text)

The description of the error that occurred for a specific import task.