Amazonka.Discovery.Types.ExportInfo

Description

Documentation

data ExportInfo

Information regarding the export status of discovered data. The value is an array of objects.

See: newExportInfo smart constructor.

Constructors

ExportInfo'

Fields

configurationsDownloadUrl :: Maybe Text

A URL for an Amazon S3 bucket where you can review the exported data. The URL is displayed only if the export succeeded.

requestedStartTime :: Maybe POSIX

The value of startTime parameter in the StartExportTask request. If no startTime was requested, this result does not appear in ExportInfo.

requestedEndTime :: Maybe POSIX

The endTime used in the StartExportTask request. If no endTime was requested, this result does not appear in ExportInfo.

isTruncated :: Maybe Bool

If true, the export of agent information exceeded the size limit for a single export and the exported data is incomplete for the requested time range. To address this, select a smaller time range for the export by using startDate and endDate.

exportId :: Text

A unique identifier used to query an export.

exportStatus :: ExportStatus

The status of the data export job.

statusMessage :: Text

A status message provided for API callers.

exportRequestTime :: POSIX

The time that the data export was initiated.

newExportInfo

Arguments

:: Text
$sel:exportId:ExportInfo'

-> ExportStatus
$sel:exportStatus:ExportInfo'

-> Text
$sel:statusMessage:ExportInfo'

-> UTCTime
$sel:exportRequestTime:ExportInfo'

-> ExportInfo

Create a value of ExportInfo with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:configurationsDownloadUrl:ExportInfo', exportInfo_configurationsDownloadUrl - A URL for an Amazon S3 bucket where you can review the exported data. The URL is displayed only if the export succeeded.

$sel:requestedStartTime:ExportInfo', exportInfo_requestedStartTime - The value of startTime parameter in the StartExportTask request. If no startTime was requested, this result does not appear in ExportInfo.

$sel:requestedEndTime:ExportInfo', exportInfo_requestedEndTime - The endTime used in the StartExportTask request. If no endTime was requested, this result does not appear in ExportInfo.

$sel:isTruncated:ExportInfo', exportInfo_isTruncated - If true, the export of agent information exceeded the size limit for a single export and the exported data is incomplete for the requested time range. To address this, select a smaller time range for the export by using startDate and endDate.

$sel:exportId:ExportInfo', exportInfo_exportId - A unique identifier used to query an export.

$sel:exportStatus:ExportInfo', exportInfo_exportStatus - The status of the data export job.

$sel:statusMessage:ExportInfo', exportInfo_statusMessage - A status message provided for API callers.

$sel:exportRequestTime:ExportInfo', exportInfo_exportRequestTime - The time that the data export was initiated.

exportInfo_configurationsDownloadUrl :: Lens' ExportInfo (Maybe Text)

A URL for an Amazon S3 bucket where you can review the exported data. The URL is displayed only if the export succeeded.

exportInfo_requestedStartTime :: Lens' ExportInfo (Maybe UTCTime)

The value of startTime parameter in the StartExportTask request. If no startTime was requested, this result does not appear in ExportInfo.

exportInfo_requestedEndTime :: Lens' ExportInfo (Maybe UTCTime)

The endTime used in the StartExportTask request. If no endTime was requested, this result does not appear in ExportInfo.

exportInfo_isTruncated :: Lens' ExportInfo (Maybe Bool)

If true, the export of agent information exceeded the size limit for a single export and the exported data is incomplete for the requested time range. To address this, select a smaller time range for the export by using startDate and endDate.

exportInfo_exportId :: Lens' ExportInfo Text

A unique identifier used to query an export.

exportInfo_exportStatus :: Lens' ExportInfo ExportStatus

The status of the data export job.

exportInfo_statusMessage :: Lens' ExportInfo Text

A status message provided for API callers.

exportInfo_exportRequestTime :: Lens' ExportInfo UTCTime

The time that the data export was initiated.