Amazonka.Discovery.Types.ImportTask

Description

Documentation

data ImportTask

An array of information related to the import task request that includes status information, times, IDs, the Amazon S3 Object URL for the import file, and more.

See: newImportTask smart constructor.

Constructors

ImportTask'

Fields

applicationImportSuccess :: Maybe Int

The total number of application records in the import file that were successfully imported.

status :: Maybe ImportStatus

The status of the import task. An import can have the status of IMPORT_COMPLETE and still have some records fail to import from the overall request. More information can be found in the downloadable archive defined in the errorsAndFailedEntriesZip field, or in the Migration Hub management console.

serverImportSuccess :: Maybe Int

The total number of server records in the import file that were successfully imported.

importCompletionTime :: Maybe POSIX

The time that the import task request finished, presented in the Unix time stamp format.

name :: Maybe Text

A descriptive name for an import task. You can use this name to filter future requests related to this import task, such as identifying applications and servers that were included in this import task. We recommend that you use a meaningful name for each import task.

applicationImportFailure :: Maybe Int

The total number of application records in the import file that failed to be imported.

errorsAndFailedEntriesZip :: Maybe Text

A link to a compressed archive folder (in the ZIP format) that contains an error log and a file of failed records. You can use these two files to quickly identify records that failed, why they failed, and correct those records. Afterward, you can upload the corrected file to your Amazon S3 bucket and create another import task request.

This field also includes authorization information so you can confirm the authenticity of the compressed archive before you download it.

If some records failed to be imported we recommend that you correct the records in the failed entries file and then imports that failed entries file. This prevents you from having to correct and update the larger original file and attempt importing it again.

importTaskId :: Maybe Text

The unique ID for a specific import task. These IDs aren't globally unique, but they are unique within an AWS account.

importDeletedTime :: Maybe POSIX

The time that the import task request was deleted, presented in the Unix time stamp format.

serverImportFailure :: Maybe Int

The total number of server records in the import file that failed to be imported.

clientRequestToken :: Maybe Text

A unique token used to prevent the same import request from occurring more than once. If you didn't provide a token, a token was automatically generated when the import task request was sent.

importUrl :: Maybe Text

The URL for your import file that you've uploaded to Amazon S3.

importRequestTime :: Maybe POSIX

The time that the import task request was made, presented in the Unix time stamp format.

newImportTask :: ImportTask

Create a value of ImportTask with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:applicationImportSuccess:ImportTask', importTask_applicationImportSuccess - The total number of application records in the import file that were successfully imported.

$sel:status:ImportTask', importTask_status - The status of the import task. An import can have the status of IMPORT_COMPLETE and still have some records fail to import from the overall request. More information can be found in the downloadable archive defined in the errorsAndFailedEntriesZip field, or in the Migration Hub management console.

$sel:serverImportSuccess:ImportTask', importTask_serverImportSuccess - The total number of server records in the import file that were successfully imported.

$sel:importCompletionTime:ImportTask', importTask_importCompletionTime - The time that the import task request finished, presented in the Unix time stamp format.

$sel:name:ImportTask', importTask_name - A descriptive name for an import task. You can use this name to filter future requests related to this import task, such as identifying applications and servers that were included in this import task. We recommend that you use a meaningful name for each import task.

$sel:applicationImportFailure:ImportTask', importTask_applicationImportFailure - The total number of application records in the import file that failed to be imported.

$sel:errorsAndFailedEntriesZip:ImportTask', importTask_errorsAndFailedEntriesZip - A link to a compressed archive folder (in the ZIP format) that contains an error log and a file of failed records. You can use these two files to quickly identify records that failed, why they failed, and correct those records. Afterward, you can upload the corrected file to your Amazon S3 bucket and create another import task request.

This field also includes authorization information so you can confirm the authenticity of the compressed archive before you download it.

If some records failed to be imported we recommend that you correct the records in the failed entries file and then imports that failed entries file. This prevents you from having to correct and update the larger original file and attempt importing it again.

$sel:importTaskId:ImportTask', importTask_importTaskId - The unique ID for a specific import task. These IDs aren't globally unique, but they are unique within an AWS account.

$sel:importDeletedTime:ImportTask', importTask_importDeletedTime - The time that the import task request was deleted, presented in the Unix time stamp format.

$sel:serverImportFailure:ImportTask', importTask_serverImportFailure - The total number of server records in the import file that failed to be imported.

$sel:clientRequestToken:ImportTask', importTask_clientRequestToken - A unique token used to prevent the same import request from occurring more than once. If you didn't provide a token, a token was automatically generated when the import task request was sent.

$sel:importUrl:ImportTask', importTask_importUrl - The URL for your import file that you've uploaded to Amazon S3.

$sel:importRequestTime:ImportTask', importTask_importRequestTime - The time that the import task request was made, presented in the Unix time stamp format.

importTask_applicationImportSuccess :: Lens' ImportTask (Maybe Int)

The total number of application records in the import file that were successfully imported.

importTask_status :: Lens' ImportTask (Maybe ImportStatus)

The status of the import task. An import can have the status of IMPORT_COMPLETE and still have some records fail to import from the overall request. More information can be found in the downloadable archive defined in the errorsAndFailedEntriesZip field, or in the Migration Hub management console.

importTask_serverImportSuccess :: Lens' ImportTask (Maybe Int)

The total number of server records in the import file that were successfully imported.

importTask_importCompletionTime :: Lens' ImportTask (Maybe UTCTime)

The time that the import task request finished, presented in the Unix time stamp format.

importTask_name :: Lens' ImportTask (Maybe Text)

A descriptive name for an import task. You can use this name to filter future requests related to this import task, such as identifying applications and servers that were included in this import task. We recommend that you use a meaningful name for each import task.

importTask_applicationImportFailure :: Lens' ImportTask (Maybe Int)

The total number of application records in the import file that failed to be imported.

importTask_errorsAndFailedEntriesZip :: Lens' ImportTask (Maybe Text)

A link to a compressed archive folder (in the ZIP format) that contains an error log and a file of failed records. You can use these two files to quickly identify records that failed, why they failed, and correct those records. Afterward, you can upload the corrected file to your Amazon S3 bucket and create another import task request.

This field also includes authorization information so you can confirm the authenticity of the compressed archive before you download it.

If some records failed to be imported we recommend that you correct the records in the failed entries file and then imports that failed entries file. This prevents you from having to correct and update the larger original file and attempt importing it again.

importTask_importTaskId :: Lens' ImportTask (Maybe Text)

The unique ID for a specific import task. These IDs aren't globally unique, but they are unique within an AWS account.

importTask_importDeletedTime :: Lens' ImportTask (Maybe UTCTime)

The time that the import task request was deleted, presented in the Unix time stamp format.

importTask_serverImportFailure :: Lens' ImportTask (Maybe Int)

The total number of server records in the import file that failed to be imported.

importTask_clientRequestToken :: Lens' ImportTask (Maybe Text)

A unique token used to prevent the same import request from occurring more than once. If you didn't provide a token, a token was automatically generated when the import task request was sent.

importTask_importUrl :: Lens' ImportTask (Maybe Text)

The URL for your import file that you've uploaded to Amazon S3.

importTask_importRequestTime :: Lens' ImportTask (Maybe UTCTime)

The time that the import task request was made, presented in the Unix time stamp format.