Amazonka.DocumentDB.CreateDBClusterParameterGroup

Description

Creates a new cluster parameter group.

Parameters in a cluster parameter group apply to all of the instances in a cluster.

A cluster parameter group is initially created with the default parameters for the database engine used by instances in the cluster. In Amazon DocumentDB, you cannot make modifications directly to the default.docdb3.6 cluster parameter group. If your Amazon DocumentDB cluster is using the default cluster parameter group and you want to modify a value in it, you must first create a new parameter group or copy an existing parameter group, modify it, and then apply the modified parameter group to your cluster. For the new cluster parameter group and associated settings to take effect, you must then reboot the instances in the cluster without failover. For more information, see Modifying Amazon DocumentDB Cluster Parameter Groups.

Creating a Request

data CreateDBClusterParameterGroup

Represents the input of CreateDBClusterParameterGroup.

See: newCreateDBClusterParameterGroup smart constructor.

Constructors

CreateDBClusterParameterGroup'

Fields

tags :: Maybe [Tag]

The tags to be assigned to the cluster parameter group.

dbClusterParameterGroupName :: Text

The name of the cluster parameter group.

Constraints:

Must not match the name of an existing DBClusterParameterGroup.

This value is stored as a lowercase string.

dbParameterGroupFamily :: Text

The cluster parameter group family name.

description :: Text

The description for the cluster parameter group.

newCreateDBClusterParameterGroup

Arguments

:: Text
$sel:dbClusterParameterGroupName:CreateDBClusterParameterGroup'

-> Text
$sel:dbParameterGroupFamily:CreateDBClusterParameterGroup'

-> Text
$sel:description:CreateDBClusterParameterGroup'

-> CreateDBClusterParameterGroup

Create a value of CreateDBClusterParameterGroup with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:tags:CreateDBClusterParameterGroup', createDBClusterParameterGroup_tags - The tags to be assigned to the cluster parameter group.

$sel:dbClusterParameterGroupName:CreateDBClusterParameterGroup', createDBClusterParameterGroup_dbClusterParameterGroupName - The name of the cluster parameter group.

Constraints:

Must not match the name of an existing DBClusterParameterGroup.

This value is stored as a lowercase string.

$sel:dbParameterGroupFamily:CreateDBClusterParameterGroup', createDBClusterParameterGroup_dbParameterGroupFamily - The cluster parameter group family name.

$sel:description:CreateDBClusterParameterGroup', createDBClusterParameterGroup_description - The description for the cluster parameter group.

Request Lenses

createDBClusterParameterGroup_tags :: Lens' CreateDBClusterParameterGroup (Maybe [Tag])

The tags to be assigned to the cluster parameter group.

createDBClusterParameterGroup_dbClusterParameterGroupName :: Lens' CreateDBClusterParameterGroup Text

The name of the cluster parameter group.

Constraints:

Must not match the name of an existing DBClusterParameterGroup.

This value is stored as a lowercase string.

createDBClusterParameterGroup_dbParameterGroupFamily :: Lens' CreateDBClusterParameterGroup Text

The cluster parameter group family name.

createDBClusterParameterGroup_description :: Lens' CreateDBClusterParameterGroup Text

The description for the cluster parameter group.

Destructuring the Response

data CreateDBClusterParameterGroupResponse

See: newCreateDBClusterParameterGroupResponse smart constructor.

Constructors

CreateDBClusterParameterGroupResponse'

Fields

dbClusterParameterGroup :: Maybe DBClusterParameterGroup

httpStatus :: Int

The response's http status code.

newCreateDBClusterParameterGroupResponse

Arguments

:: Int
$sel:httpStatus:CreateDBClusterParameterGroupResponse'

-> CreateDBClusterParameterGroupResponse

Create a value of CreateDBClusterParameterGroupResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:dbClusterParameterGroup:CreateDBClusterParameterGroupResponse', createDBClusterParameterGroupResponse_dbClusterParameterGroup - Undocumented member.

$sel:httpStatus:CreateDBClusterParameterGroupResponse', createDBClusterParameterGroupResponse_httpStatus - The response's http status code.

Response Lenses

createDBClusterParameterGroupResponse_dbClusterParameterGroup :: Lens' CreateDBClusterParameterGroupResponse (Maybe DBClusterParameterGroup)

Undocumented member.

createDBClusterParameterGroupResponse_httpStatus :: Lens' CreateDBClusterParameterGroupResponse Int

The response's http status code.