Amazonka.DocumentDB.CreateDBInstance

Description

Creates a new instance.

Creating a Request

data CreateDBInstance

Represents the input to CreateDBInstance.

See: newCreateDBInstance smart constructor.

Constructors

CreateDBInstance'

Fields

autoMinorVersionUpgrade :: Maybe Bool

This parameter does not apply to Amazon DocumentDB. Amazon DocumentDB does not perform minor version upgrades regardless of the value set.

Default: false

promotionTier :: Maybe Int

A value that specifies the order in which an Amazon DocumentDB replica is promoted to the primary instance after a failure of the existing primary instance.

Default: 1

Valid values: 0-15

preferredMaintenanceWindow :: Maybe Text

The time range each week during which system maintenance can occur, in Universal Coordinated Time (UTC).

Format: ddd:hh24:mi-ddd:hh24:mi

The default is a 30-minute window selected at random from an 8-hour block of time for each Region, occurring on a random day of the week.

Valid days: Mon, Tue, Wed, Thu, Fri, Sat, Sun

Constraints: Minimum 30-minute window.

availabilityZone :: Maybe Text

The Amazon EC2 Availability Zone that the instance is created in.

Default: A random, system-chosen Availability Zone in the endpoint's Region.

Example: us-east-1d

tags :: Maybe [Tag]

The tags to be assigned to the instance. You can assign up to 10 tags to an instance.

dbInstanceIdentifier :: Text

The instance identifier. This parameter is stored as a lowercase string.

Constraints:

Must contain from 1 to 63 letters, numbers, or hyphens.

The first character must be a letter.

Cannot end with a hyphen or contain two consecutive hyphens.

Example: mydbinstance

dbInstanceClass :: Text

The compute and memory capacity of the instance; for example, db.r5.large.

engine :: Text

The name of the database engine to be used for this instance.

Valid value: docdb

dbClusterIdentifier :: Text

The identifier of the cluster that the instance will belong to.

newCreateDBInstance

Arguments

:: Text
$sel:dbInstanceIdentifier:CreateDBInstance'

-> Text
$sel:dbInstanceClass:CreateDBInstance'

-> Text
$sel:engine:CreateDBInstance'

-> Text
$sel:dbClusterIdentifier:CreateDBInstance'

-> CreateDBInstance

Create a value of CreateDBInstance with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:autoMinorVersionUpgrade:CreateDBInstance', createDBInstance_autoMinorVersionUpgrade - This parameter does not apply to Amazon DocumentDB. Amazon DocumentDB does not perform minor version upgrades regardless of the value set.

Default: false

$sel:promotionTier:CreateDBInstance', createDBInstance_promotionTier - A value that specifies the order in which an Amazon DocumentDB replica is promoted to the primary instance after a failure of the existing primary instance.

Default: 1

Valid values: 0-15

$sel:preferredMaintenanceWindow:CreateDBInstance', createDBInstance_preferredMaintenanceWindow - The time range each week during which system maintenance can occur, in Universal Coordinated Time (UTC).

Format: ddd:hh24:mi-ddd:hh24:mi

The default is a 30-minute window selected at random from an 8-hour block of time for each Region, occurring on a random day of the week.

Valid days: Mon, Tue, Wed, Thu, Fri, Sat, Sun

Constraints: Minimum 30-minute window.

$sel:availabilityZone:CreateDBInstance', createDBInstance_availabilityZone - The Amazon EC2 Availability Zone that the instance is created in.

Default: A random, system-chosen Availability Zone in the endpoint's Region.

Example: us-east-1d

$sel:tags:CreateDBInstance', createDBInstance_tags - The tags to be assigned to the instance. You can assign up to 10 tags to an instance.

$sel:dbInstanceIdentifier:CreateDBInstance', createDBInstance_dbInstanceIdentifier - The instance identifier. This parameter is stored as a lowercase string.

Constraints:

Must contain from 1 to 63 letters, numbers, or hyphens.

The first character must be a letter.

Cannot end with a hyphen or contain two consecutive hyphens.

Example: mydbinstance

$sel:dbInstanceClass:CreateDBInstance', createDBInstance_dbInstanceClass - The compute and memory capacity of the instance; for example, db.r5.large.

$sel:engine:CreateDBInstance', createDBInstance_engine - The name of the database engine to be used for this instance.

Valid value: docdb

$sel:dbClusterIdentifier:CreateDBInstance', createDBInstance_dbClusterIdentifier - The identifier of the cluster that the instance will belong to.

Request Lenses

createDBInstance_autoMinorVersionUpgrade :: Lens' CreateDBInstance (Maybe Bool)

This parameter does not apply to Amazon DocumentDB. Amazon DocumentDB does not perform minor version upgrades regardless of the value set.

Default: false

createDBInstance_promotionTier :: Lens' CreateDBInstance (Maybe Int)

A value that specifies the order in which an Amazon DocumentDB replica is promoted to the primary instance after a failure of the existing primary instance.

Default: 1

Valid values: 0-15

createDBInstance_preferredMaintenanceWindow :: Lens' CreateDBInstance (Maybe Text)

The time range each week during which system maintenance can occur, in Universal Coordinated Time (UTC).

Format: ddd:hh24:mi-ddd:hh24:mi

The default is a 30-minute window selected at random from an 8-hour block of time for each Region, occurring on a random day of the week.

Valid days: Mon, Tue, Wed, Thu, Fri, Sat, Sun

Constraints: Minimum 30-minute window.

createDBInstance_availabilityZone :: Lens' CreateDBInstance (Maybe Text)

The Amazon EC2 Availability Zone that the instance is created in.

Default: A random, system-chosen Availability Zone in the endpoint's Region.

Example: us-east-1d

createDBInstance_tags :: Lens' CreateDBInstance (Maybe [Tag])

The tags to be assigned to the instance. You can assign up to 10 tags to an instance.

createDBInstance_dbInstanceIdentifier :: Lens' CreateDBInstance Text

The instance identifier. This parameter is stored as a lowercase string.

Constraints:

Must contain from 1 to 63 letters, numbers, or hyphens.

The first character must be a letter.

Cannot end with a hyphen or contain two consecutive hyphens.

Example: mydbinstance

createDBInstance_dbInstanceClass :: Lens' CreateDBInstance Text

The compute and memory capacity of the instance; for example, db.r5.large.

createDBInstance_engine :: Lens' CreateDBInstance Text

The name of the database engine to be used for this instance.

Valid value: docdb

createDBInstance_dbClusterIdentifier :: Lens' CreateDBInstance Text

The identifier of the cluster that the instance will belong to.

Destructuring the Response

data CreateDBInstanceResponse

See: newCreateDBInstanceResponse smart constructor.

Constructors

CreateDBInstanceResponse'

Fields

dbInstance :: Maybe DBInstance

httpStatus :: Int

The response's http status code.

newCreateDBInstanceResponse

Arguments

:: Int
$sel:httpStatus:CreateDBInstanceResponse'

-> CreateDBInstanceResponse

Create a value of CreateDBInstanceResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:dbInstance:CreateDBInstanceResponse', createDBInstanceResponse_dbInstance - Undocumented member.

$sel:httpStatus:CreateDBInstanceResponse', createDBInstanceResponse_httpStatus - The response's http status code.

Response Lenses

createDBInstanceResponse_dbInstance :: Lens' CreateDBInstanceResponse (Maybe DBInstance)

Undocumented member.

createDBInstanceResponse_httpStatus :: Lens' CreateDBInstanceResponse Int

The response's http status code.