Amazonka.DocumentDB.CreateGlobalCluster

Description

Creates an Amazon DocumentDB global cluster that can span multiple multiple Regions. The global cluster contains one primary cluster with read-write capability, and up-to give read-only secondary clusters. Global clusters uses storage-based fast replication across regions with latencies less than one second, using dedicated infrastructure with no impact to your workload’s performance.

You can create a global cluster that is initially empty, and then add a primary and a secondary to it. Or you can specify an existing cluster during the create operation, and this cluster becomes the primary of the global cluster.

This action only applies to Amazon DocumentDB clusters.

Creating a Request

data CreateGlobalCluster

Represents the input to CreateGlobalCluster.

See: newCreateGlobalCluster smart constructor.

Constructors

CreateGlobalCluster'

Fields

engineVersion :: Maybe Text

The engine version of the global cluster.

deletionProtection :: Maybe Bool

The deletion protection setting for the new global cluster. The global cluster can't be deleted when deletion protection is enabled.

storageEncrypted :: Maybe Bool

The storage encryption setting for the new global cluster.

sourceDBClusterIdentifier :: Maybe Text

The Amazon Resource Name (ARN) to use as the primary cluster of the global cluster. This parameter is optional.

engine :: Maybe Text

The name of the database engine to be used for this cluster.

databaseName :: Maybe Text

The name for your database of up to 64 alpha-numeric characters. If you do not provide a name, Amazon DocumentDB will not create a database in the global cluster you are creating.

globalClusterIdentifier :: Text

The cluster identifier of the new global cluster.

newCreateGlobalCluster

Arguments

:: Text
$sel:globalClusterIdentifier:CreateGlobalCluster'

-> CreateGlobalCluster

Create a value of CreateGlobalCluster with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:engineVersion:CreateGlobalCluster', createGlobalCluster_engineVersion - The engine version of the global cluster.

$sel:deletionProtection:CreateGlobalCluster', createGlobalCluster_deletionProtection - The deletion protection setting for the new global cluster. The global cluster can't be deleted when deletion protection is enabled.

$sel:storageEncrypted:CreateGlobalCluster', createGlobalCluster_storageEncrypted - The storage encryption setting for the new global cluster.

$sel:sourceDBClusterIdentifier:CreateGlobalCluster', createGlobalCluster_sourceDBClusterIdentifier - The Amazon Resource Name (ARN) to use as the primary cluster of the global cluster. This parameter is optional.

$sel:engine:CreateGlobalCluster', createGlobalCluster_engine - The name of the database engine to be used for this cluster.

$sel:databaseName:CreateGlobalCluster', createGlobalCluster_databaseName - The name for your database of up to 64 alpha-numeric characters. If you do not provide a name, Amazon DocumentDB will not create a database in the global cluster you are creating.

$sel:globalClusterIdentifier:CreateGlobalCluster', createGlobalCluster_globalClusterIdentifier - The cluster identifier of the new global cluster.

Request Lenses

createGlobalCluster_engineVersion :: Lens' CreateGlobalCluster (Maybe Text)

The engine version of the global cluster.

createGlobalCluster_deletionProtection :: Lens' CreateGlobalCluster (Maybe Bool)

The deletion protection setting for the new global cluster. The global cluster can't be deleted when deletion protection is enabled.

createGlobalCluster_storageEncrypted :: Lens' CreateGlobalCluster (Maybe Bool)

The storage encryption setting for the new global cluster.

createGlobalCluster_sourceDBClusterIdentifier :: Lens' CreateGlobalCluster (Maybe Text)

The Amazon Resource Name (ARN) to use as the primary cluster of the global cluster. This parameter is optional.

createGlobalCluster_engine :: Lens' CreateGlobalCluster (Maybe Text)

The name of the database engine to be used for this cluster.

createGlobalCluster_databaseName :: Lens' CreateGlobalCluster (Maybe Text)

The name for your database of up to 64 alpha-numeric characters. If you do not provide a name, Amazon DocumentDB will not create a database in the global cluster you are creating.

createGlobalCluster_globalClusterIdentifier :: Lens' CreateGlobalCluster Text

The cluster identifier of the new global cluster.

Destructuring the Response

data CreateGlobalClusterResponse

See: newCreateGlobalClusterResponse smart constructor.

Constructors

CreateGlobalClusterResponse'

Fields

globalCluster :: Maybe GlobalCluster

httpStatus :: Int

The response's http status code.

newCreateGlobalClusterResponse

Arguments

:: Int
$sel:httpStatus:CreateGlobalClusterResponse'

-> CreateGlobalClusterResponse

Create a value of CreateGlobalClusterResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:globalCluster:CreateGlobalClusterResponse', createGlobalClusterResponse_globalCluster - Undocumented member.

$sel:httpStatus:CreateGlobalClusterResponse', createGlobalClusterResponse_httpStatus - The response's http status code.

Response Lenses

createGlobalClusterResponse_globalCluster :: Lens' CreateGlobalClusterResponse (Maybe GlobalCluster)

Undocumented member.

createGlobalClusterResponse_httpStatus :: Lens' CreateGlobalClusterResponse Int

The response's http status code.