Amazonka.DocumentDB.DescribeDBSubnetGroups

Description

Returns a list of DBSubnetGroup descriptions. If a DBSubnetGroupName is specified, the list will contain only the descriptions of the specified DBSubnetGroup.

This operation returns paginated results.

Creating a Request

data DescribeDBSubnetGroups

Represents the input to DescribeDBSubnetGroups.

See: newDescribeDBSubnetGroups smart constructor.

Constructors

DescribeDBSubnetGroups'

Fields

dbSubnetGroupName :: Maybe Text

The name of the subnet group to return details for.

filters :: Maybe [Filter]

This parameter is not currently supported.

marker :: Maybe Text

An optional pagination token provided by a previous request. If this parameter is specified, the response includes only records beyond the marker, up to the value specified by MaxRecords.

maxRecords :: Maybe Int

The maximum number of records to include in the response. If more records exist than the specified MaxRecords value, a pagination token (marker) is included in the response so that the remaining results can be retrieved.

Default: 100

Constraints: Minimum 20, maximum 100.

newDescribeDBSubnetGroups :: DescribeDBSubnetGroups

Create a value of DescribeDBSubnetGroups with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:dbSubnetGroupName:DescribeDBSubnetGroups', describeDBSubnetGroups_dbSubnetGroupName - The name of the subnet group to return details for.

$sel:filters:DescribeDBSubnetGroups', describeDBSubnetGroups_filters - This parameter is not currently supported.

$sel:marker:DescribeDBSubnetGroups', describeDBSubnetGroups_marker - An optional pagination token provided by a previous request. If this parameter is specified, the response includes only records beyond the marker, up to the value specified by MaxRecords.

$sel:maxRecords:DescribeDBSubnetGroups', describeDBSubnetGroups_maxRecords - The maximum number of records to include in the response. If more records exist than the specified MaxRecords value, a pagination token (marker) is included in the response so that the remaining results can be retrieved.

Default: 100

Constraints: Minimum 20, maximum 100.

Request Lenses

describeDBSubnetGroups_dbSubnetGroupName :: Lens' DescribeDBSubnetGroups (Maybe Text)

The name of the subnet group to return details for.

describeDBSubnetGroups_filters :: Lens' DescribeDBSubnetGroups (Maybe [Filter])

This parameter is not currently supported.

describeDBSubnetGroups_marker :: Lens' DescribeDBSubnetGroups (Maybe Text)

An optional pagination token provided by a previous request. If this parameter is specified, the response includes only records beyond the marker, up to the value specified by MaxRecords.

describeDBSubnetGroups_maxRecords :: Lens' DescribeDBSubnetGroups (Maybe Int)

The maximum number of records to include in the response. If more records exist than the specified MaxRecords value, a pagination token (marker) is included in the response so that the remaining results can be retrieved.

Default: 100

Constraints: Minimum 20, maximum 100.

Destructuring the Response

data DescribeDBSubnetGroupsResponse

Represents the output of DescribeDBSubnetGroups.

See: newDescribeDBSubnetGroupsResponse smart constructor.

Constructors

DescribeDBSubnetGroupsResponse'

Fields

dbSubnetGroups :: Maybe [DBSubnetGroup]

Detailed information about one or more subnet groups.

marker :: Maybe Text

An optional pagination token provided by a previous request. If this parameter is specified, the response includes only records beyond the marker, up to the value specified by MaxRecords.

httpStatus :: Int

The response's http status code.

newDescribeDBSubnetGroupsResponse

Arguments

:: Int
$sel:httpStatus:DescribeDBSubnetGroupsResponse'

-> DescribeDBSubnetGroupsResponse

Create a value of DescribeDBSubnetGroupsResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:dbSubnetGroups:DescribeDBSubnetGroupsResponse', describeDBSubnetGroupsResponse_dbSubnetGroups - Detailed information about one or more subnet groups.

$sel:marker:DescribeDBSubnetGroups', describeDBSubnetGroupsResponse_marker - An optional pagination token provided by a previous request. If this parameter is specified, the response includes only records beyond the marker, up to the value specified by MaxRecords.

$sel:httpStatus:DescribeDBSubnetGroupsResponse', describeDBSubnetGroupsResponse_httpStatus - The response's http status code.

Response Lenses

describeDBSubnetGroupsResponse_dbSubnetGroups :: Lens' DescribeDBSubnetGroupsResponse (Maybe [DBSubnetGroup])

Detailed information about one or more subnet groups.

describeDBSubnetGroupsResponse_marker :: Lens' DescribeDBSubnetGroupsResponse (Maybe Text)

An optional pagination token provided by a previous request. If this parameter is specified, the response includes only records beyond the marker, up to the value specified by MaxRecords.

describeDBSubnetGroupsResponse_httpStatus :: Lens' DescribeDBSubnetGroupsResponse Int

The response's http status code.