Amazonka.DocumentDB.ModifyDBCluster

Description

Modifies a setting for an Amazon DocumentDB cluster. You can change one or more database configuration parameters by specifying these parameters and the new values in the request.

Creating a Request

data ModifyDBCluster

Represents the input to ModifyDBCluster.

See: newModifyDBCluster smart constructor.

Constructors

ModifyDBCluster'

Fields

engineVersion :: Maybe Text

The version number of the database engine to which you want to upgrade. Modifying engine version is not supported on Amazon DocumentDB.

deletionProtection :: Maybe Bool

Specifies whether this cluster can be deleted. If DeletionProtection is enabled, the cluster cannot be deleted unless it is modified and DeletionProtection is disabled. DeletionProtection protects clusters from being accidentally deleted.

masterUserPassword :: Maybe Text

The password for the master database user. This password can contain any printable ASCII character except forward slash (/), double quote ("), or the "at" symbol (@).

Constraints: Must contain from 8 to 100 characters.

cloudwatchLogsExportConfiguration :: Maybe CloudwatchLogsExportConfiguration

The configuration setting for the log types to be enabled for export to Amazon CloudWatch Logs for a specific instance or cluster. The EnableLogTypes and DisableLogTypes arrays determine which logs are exported (or not exported) to CloudWatch Logs.

preferredMaintenanceWindow :: Maybe Text

The weekly time range during which system maintenance can occur, in Universal Coordinated Time (UTC).

Format: ddd:hh24:mi-ddd:hh24:mi

The default is a 30-minute window selected at random from an 8-hour block of time for each Region, occurring on a random day of the week.

Valid days: Mon, Tue, Wed, Thu, Fri, Sat, Sun

Constraints: Minimum 30-minute window.

preferredBackupWindow :: Maybe Text

The daily time range during which automated backups are created if automated backups are enabled, using the BackupRetentionPeriod parameter.

The default is a 30-minute window selected at random from an 8-hour block of time for each Region.

Constraints:

Must be in the format hh24:mi-hh24:mi.

Must be in Universal Coordinated Time (UTC).

Must not conflict with the preferred maintenance window.

Must be at least 30 minutes.

backupRetentionPeriod :: Maybe Int

The number of days for which automated backups are retained. You must specify a minimum value of 1.

Default: 1

Constraints:

Must be a value from 1 to 35.

vpcSecurityGroupIds :: Maybe [Text]

A list of virtual private cloud (VPC) security groups that the cluster will belong to.

dbClusterParameterGroupName :: Maybe Text

The name of the cluster parameter group to use for the cluster.

applyImmediately :: Maybe Bool

A value that specifies whether the changes in this request and any pending changes are asynchronously applied as soon as possible, regardless of the PreferredMaintenanceWindow setting for the cluster. If this parameter is set to false, changes to the cluster are applied during the next maintenance window.

The ApplyImmediately parameter affects only the NewDBClusterIdentifier and MasterUserPassword values. If you set this parameter value to false, the changes to the NewDBClusterIdentifier and MasterUserPassword values are applied during the next maintenance window. All other changes are applied immediately, regardless of the value of the ApplyImmediately parameter.

Default: false

newDBClusterIdentifier' :: Maybe Text

The new cluster identifier for the cluster when renaming a cluster. This value is stored as a lowercase string.

Constraints:

Must contain from 1 to 63 letters, numbers, or hyphens.

The first character must be a letter.

Cannot end with a hyphen or contain two consecutive hyphens.

Example: my-cluster2

port :: Maybe Int

The port number on which the cluster accepts connections.

Constraints: Must be a value from 1150 to 65535.

Default: The same port as the original cluster.

dbClusterIdentifier :: Text

The cluster identifier for the cluster that is being modified. This parameter is not case sensitive.

Constraints:

Must match the identifier of an existing DBCluster.

newModifyDBCluster

Arguments

:: Text
$sel:dbClusterIdentifier:ModifyDBCluster'

-> ModifyDBCluster

Create a value of ModifyDBCluster with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:engineVersion:ModifyDBCluster', modifyDBCluster_engineVersion - The version number of the database engine to which you want to upgrade. Modifying engine version is not supported on Amazon DocumentDB.

$sel:deletionProtection:ModifyDBCluster', modifyDBCluster_deletionProtection - Specifies whether this cluster can be deleted. If DeletionProtection is enabled, the cluster cannot be deleted unless it is modified and DeletionProtection is disabled. DeletionProtection protects clusters from being accidentally deleted.

$sel:masterUserPassword:ModifyDBCluster', modifyDBCluster_masterUserPassword - The password for the master database user. This password can contain any printable ASCII character except forward slash (/), double quote ("), or the "at" symbol (@).

Constraints: Must contain from 8 to 100 characters.

$sel:cloudwatchLogsExportConfiguration:ModifyDBCluster', modifyDBCluster_cloudwatchLogsExportConfiguration - The configuration setting for the log types to be enabled for export to Amazon CloudWatch Logs for a specific instance or cluster. The EnableLogTypes and DisableLogTypes arrays determine which logs are exported (or not exported) to CloudWatch Logs.

$sel:preferredMaintenanceWindow:ModifyDBCluster', modifyDBCluster_preferredMaintenanceWindow - The weekly time range during which system maintenance can occur, in Universal Coordinated Time (UTC).

Format: ddd:hh24:mi-ddd:hh24:mi

The default is a 30-minute window selected at random from an 8-hour block of time for each Region, occurring on a random day of the week.

Valid days: Mon, Tue, Wed, Thu, Fri, Sat, Sun

Constraints: Minimum 30-minute window.

$sel:preferredBackupWindow:ModifyDBCluster', modifyDBCluster_preferredBackupWindow - The daily time range during which automated backups are created if automated backups are enabled, using the BackupRetentionPeriod parameter.

The default is a 30-minute window selected at random from an 8-hour block of time for each Region.

Constraints:

Must be in the format hh24:mi-hh24:mi.

Must be in Universal Coordinated Time (UTC).

Must not conflict with the preferred maintenance window.

Must be at least 30 minutes.

$sel:backupRetentionPeriod:ModifyDBCluster', modifyDBCluster_backupRetentionPeriod - The number of days for which automated backups are retained. You must specify a minimum value of 1.

Default: 1

Constraints:

Must be a value from 1 to 35.

$sel:vpcSecurityGroupIds:ModifyDBCluster', modifyDBCluster_vpcSecurityGroupIds - A list of virtual private cloud (VPC) security groups that the cluster will belong to.

$sel:dbClusterParameterGroupName:ModifyDBCluster', modifyDBCluster_dbClusterParameterGroupName - The name of the cluster parameter group to use for the cluster.

$sel:applyImmediately:ModifyDBCluster', modifyDBCluster_applyImmediately - A value that specifies whether the changes in this request and any pending changes are asynchronously applied as soon as possible, regardless of the PreferredMaintenanceWindow setting for the cluster. If this parameter is set to false, changes to the cluster are applied during the next maintenance window.

The ApplyImmediately parameter affects only the NewDBClusterIdentifier and MasterUserPassword values. If you set this parameter value to false, the changes to the NewDBClusterIdentifier and MasterUserPassword values are applied during the next maintenance window. All other changes are applied immediately, regardless of the value of the ApplyImmediately parameter.

Default: false

$sel:newDBClusterIdentifier':ModifyDBCluster', modifyDBCluster_newDBClusterIdentifier - The new cluster identifier for the cluster when renaming a cluster. This value is stored as a lowercase string.

Constraints:

Must contain from 1 to 63 letters, numbers, or hyphens.

The first character must be a letter.

Cannot end with a hyphen or contain two consecutive hyphens.

Example: my-cluster2

$sel:port:ModifyDBCluster', modifyDBCluster_port - The port number on which the cluster accepts connections.

Constraints: Must be a value from 1150 to 65535.

Default: The same port as the original cluster.

$sel:dbClusterIdentifier:ModifyDBCluster', modifyDBCluster_dbClusterIdentifier - The cluster identifier for the cluster that is being modified. This parameter is not case sensitive.

Constraints:

Must match the identifier of an existing DBCluster.

Request Lenses

modifyDBCluster_engineVersion :: Lens' ModifyDBCluster (Maybe Text)

The version number of the database engine to which you want to upgrade. Modifying engine version is not supported on Amazon DocumentDB.

modifyDBCluster_deletionProtection :: Lens' ModifyDBCluster (Maybe Bool)

Specifies whether this cluster can be deleted. If DeletionProtection is enabled, the cluster cannot be deleted unless it is modified and DeletionProtection is disabled. DeletionProtection protects clusters from being accidentally deleted.

modifyDBCluster_masterUserPassword :: Lens' ModifyDBCluster (Maybe Text)

The password for the master database user. This password can contain any printable ASCII character except forward slash (/), double quote ("), or the "at" symbol (@).

Constraints: Must contain from 8 to 100 characters.

modifyDBCluster_cloudwatchLogsExportConfiguration :: Lens' ModifyDBCluster (Maybe CloudwatchLogsExportConfiguration)

The configuration setting for the log types to be enabled for export to Amazon CloudWatch Logs for a specific instance or cluster. The EnableLogTypes and DisableLogTypes arrays determine which logs are exported (or not exported) to CloudWatch Logs.

modifyDBCluster_preferredMaintenanceWindow :: Lens' ModifyDBCluster (Maybe Text)

The weekly time range during which system maintenance can occur, in Universal Coordinated Time (UTC).

Format: ddd:hh24:mi-ddd:hh24:mi

The default is a 30-minute window selected at random from an 8-hour block of time for each Region, occurring on a random day of the week.

Valid days: Mon, Tue, Wed, Thu, Fri, Sat, Sun

Constraints: Minimum 30-minute window.

modifyDBCluster_preferredBackupWindow :: Lens' ModifyDBCluster (Maybe Text)

The daily time range during which automated backups are created if automated backups are enabled, using the BackupRetentionPeriod parameter.

The default is a 30-minute window selected at random from an 8-hour block of time for each Region.

Constraints:

Must be in the format hh24:mi-hh24:mi.

Must be in Universal Coordinated Time (UTC).

Must not conflict with the preferred maintenance window.

Must be at least 30 minutes.

modifyDBCluster_backupRetentionPeriod :: Lens' ModifyDBCluster (Maybe Int)

The number of days for which automated backups are retained. You must specify a minimum value of 1.

Default: 1

Constraints:

Must be a value from 1 to 35.

modifyDBCluster_vpcSecurityGroupIds :: Lens' ModifyDBCluster (Maybe [Text])

A list of virtual private cloud (VPC) security groups that the cluster will belong to.

modifyDBCluster_dbClusterParameterGroupName :: Lens' ModifyDBCluster (Maybe Text)

The name of the cluster parameter group to use for the cluster.

modifyDBCluster_applyImmediately :: Lens' ModifyDBCluster (Maybe Bool)

A value that specifies whether the changes in this request and any pending changes are asynchronously applied as soon as possible, regardless of the PreferredMaintenanceWindow setting for the cluster. If this parameter is set to false, changes to the cluster are applied during the next maintenance window.

The ApplyImmediately parameter affects only the NewDBClusterIdentifier and MasterUserPassword values. If you set this parameter value to false, the changes to the NewDBClusterIdentifier and MasterUserPassword values are applied during the next maintenance window. All other changes are applied immediately, regardless of the value of the ApplyImmediately parameter.

Default: false

modifyDBCluster_newDBClusterIdentifier :: Lens' ModifyDBCluster (Maybe Text)

The new cluster identifier for the cluster when renaming a cluster. This value is stored as a lowercase string.

Constraints:

Must contain from 1 to 63 letters, numbers, or hyphens.

The first character must be a letter.

Cannot end with a hyphen or contain two consecutive hyphens.

Example: my-cluster2

modifyDBCluster_port :: Lens' ModifyDBCluster (Maybe Int)

The port number on which the cluster accepts connections.

Constraints: Must be a value from 1150 to 65535.

Default: The same port as the original cluster.

modifyDBCluster_dbClusterIdentifier :: Lens' ModifyDBCluster Text

The cluster identifier for the cluster that is being modified. This parameter is not case sensitive.

Constraints:

Must match the identifier of an existing DBCluster.

Destructuring the Response

data ModifyDBClusterResponse

See: newModifyDBClusterResponse smart constructor.

Constructors

ModifyDBClusterResponse'

Fields

dbCluster :: Maybe DBCluster

httpStatus :: Int

The response's http status code.

newModifyDBClusterResponse

Arguments

:: Int
$sel:httpStatus:ModifyDBClusterResponse'

-> ModifyDBClusterResponse

Create a value of ModifyDBClusterResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:dbCluster:ModifyDBClusterResponse', modifyDBClusterResponse_dbCluster - Undocumented member.

$sel:httpStatus:ModifyDBClusterResponse', modifyDBClusterResponse_httpStatus - The response's http status code.

Response Lenses

modifyDBClusterResponse_dbCluster :: Lens' ModifyDBClusterResponse (Maybe DBCluster)

Undocumented member.

modifyDBClusterResponse_httpStatus :: Lens' ModifyDBClusterResponse Int

The response's http status code.