Amazonka.DocumentDB.ModifyGlobalCluster

Description

Modify a setting for an Amazon DocumentDB global cluster. You can change one or more configuration parameters (for example: deletion protection), or the global cluster identifier by specifying these parameters and the new values in the request.

This action only applies to Amazon DocumentDB clusters.

Creating a Request

data ModifyGlobalCluster

Represents the input to ModifyGlobalCluster.

See: newModifyGlobalCluster smart constructor.

Constructors

ModifyGlobalCluster'

Fields

deletionProtection :: Maybe Bool

Indicates if the global cluster has deletion protection enabled. The global cluster can't be deleted when deletion protection is enabled.

newGlobalClusterIdentifier' :: Maybe Text

The new identifier for a global cluster when you modify a global cluster. This value is stored as a lowercase string.

Must contain from 1 to 63 letters, numbers, or hyphens

The first character must be a letter

Can't end with a hyphen or contain two consecutive hyphens

Example: my-cluster2

globalClusterIdentifier :: Text

The identifier for the global cluster being modified. This parameter isn't case-sensitive.

Constraints:

Must match the identifier of an existing global cluster.

newModifyGlobalCluster

Arguments

:: Text
$sel:globalClusterIdentifier:ModifyGlobalCluster'

-> ModifyGlobalCluster

Create a value of ModifyGlobalCluster with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:deletionProtection:ModifyGlobalCluster', modifyGlobalCluster_deletionProtection - Indicates if the global cluster has deletion protection enabled. The global cluster can't be deleted when deletion protection is enabled.

$sel:newGlobalClusterIdentifier':ModifyGlobalCluster', modifyGlobalCluster_newGlobalClusterIdentifier - The new identifier for a global cluster when you modify a global cluster. This value is stored as a lowercase string.

Must contain from 1 to 63 letters, numbers, or hyphens

The first character must be a letter

Can't end with a hyphen or contain two consecutive hyphens

Example: my-cluster2

$sel:globalClusterIdentifier:ModifyGlobalCluster', modifyGlobalCluster_globalClusterIdentifier - The identifier for the global cluster being modified. This parameter isn't case-sensitive.

Constraints:

Must match the identifier of an existing global cluster.

Request Lenses

modifyGlobalCluster_deletionProtection :: Lens' ModifyGlobalCluster (Maybe Bool)

Indicates if the global cluster has deletion protection enabled. The global cluster can't be deleted when deletion protection is enabled.

modifyGlobalCluster_newGlobalClusterIdentifier :: Lens' ModifyGlobalCluster (Maybe Text)

The new identifier for a global cluster when you modify a global cluster. This value is stored as a lowercase string.

Must contain from 1 to 63 letters, numbers, or hyphens

The first character must be a letter

Can't end with a hyphen or contain two consecutive hyphens

Example: my-cluster2

modifyGlobalCluster_globalClusterIdentifier :: Lens' ModifyGlobalCluster Text

The identifier for the global cluster being modified. This parameter isn't case-sensitive.

Constraints:

Must match the identifier of an existing global cluster.

Destructuring the Response

data ModifyGlobalClusterResponse

See: newModifyGlobalClusterResponse smart constructor.

Constructors

ModifyGlobalClusterResponse'

Fields

globalCluster :: Maybe GlobalCluster

httpStatus :: Int

The response's http status code.

newModifyGlobalClusterResponse

Arguments

:: Int
$sel:httpStatus:ModifyGlobalClusterResponse'

-> ModifyGlobalClusterResponse

Create a value of ModifyGlobalClusterResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:globalCluster:ModifyGlobalClusterResponse', modifyGlobalClusterResponse_globalCluster - Undocumented member.

$sel:httpStatus:ModifyGlobalClusterResponse', modifyGlobalClusterResponse_httpStatus - The response's http status code.

Response Lenses

modifyGlobalClusterResponse_globalCluster :: Lens' ModifyGlobalClusterResponse (Maybe GlobalCluster)

Undocumented member.

modifyGlobalClusterResponse_httpStatus :: Lens' ModifyGlobalClusterResponse Int

The response's http status code.