Amazonka.DocumentDB.Types.DBInstance

Description

Documentation

data DBInstance

Detailed information about an instance.

See: newDBInstance smart constructor.

Constructors

DBInstance'

Fields

engineVersion :: Maybe Text

Indicates the database engine version.

storageEncrypted :: Maybe Bool

Specifies whether or not the instance is encrypted.

dbClusterIdentifier :: Maybe Text

Contains the name of the cluster that the instance is a member of if the instance is a member of a cluster.

publiclyAccessible :: Maybe Bool

Not supported. Amazon DocumentDB does not currently support public endpoints. The value of PubliclyAccessible is always false.

autoMinorVersionUpgrade :: Maybe Bool

Does not apply. This parameter does not apply to Amazon DocumentDB. Amazon DocumentDB does not perform minor version upgrades regardless of the value set.

dbInstanceArn :: Maybe Text

The Amazon Resource Name (ARN) for the instance.

instanceCreateTime :: Maybe ISO8601

Provides the date and time that the instance was created.

engine :: Maybe Text

Provides the name of the database engine to be used for this instance.

latestRestorableTime :: Maybe ISO8601

Specifies the latest time to which a database can be restored with point-in-time restore.

dbInstanceClass :: Maybe Text

Contains the name of the compute and memory capacity class of the instance.

promotionTier :: Maybe Int

A value that specifies the order in which an Amazon DocumentDB replica is promoted to the primary instance after a failure of the existing primary instance.

preferredMaintenanceWindow :: Maybe Text

Specifies the weekly time range during which system maintenance can occur, in Universal Coordinated Time (UTC).

cACertificateIdentifier :: Maybe Text

The identifier of the CA certificate for this DB instance.

dbInstanceIdentifier :: Maybe Text

Contains a user-provided database identifier. This identifier is the unique key that identifies an instance.

kmsKeyId :: Maybe Text

If StorageEncrypted is true, the KMS key identifier for the encrypted instance.

preferredBackupWindow :: Maybe Text

Specifies the daily time range during which automated backups are created if automated backups are enabled, as determined by the BackupRetentionPeriod.

availabilityZone :: Maybe Text

Specifies the name of the Availability Zone that the instance is located in.

vpcSecurityGroups :: Maybe [VpcSecurityGroupMembership]

Provides a list of VPC security group elements that the instance belongs to.

backupRetentionPeriod :: Maybe Int

Specifies the number of days for which automatic snapshots are retained.

dbSubnetGroup :: Maybe DBSubnetGroup

Specifies information on the subnet group that is associated with the instance, including the name, description, and subnets in the subnet group.

enabledCloudwatchLogsExports :: Maybe [Text]

A list of log types that this instance is configured to export to CloudWatch Logs.

dbiResourceId :: Maybe Text

The Region-unique, immutable identifier for the instance. This identifier is found in CloudTrail log entries whenever the KMS key for the instance is accessed.

endpoint :: Maybe Endpoint

Specifies the connection endpoint.

dbInstanceStatus :: Maybe Text

Specifies the current state of this database.

pendingModifiedValues :: Maybe PendingModifiedValues

Specifies that changes to the instance are pending. This element is included only when changes are pending. Specific changes are identified by subelements.

statusInfos :: Maybe [DBInstanceStatusInfo]

The status of a read replica. If the instance is not a read replica, this is blank.

newDBInstance :: DBInstance

Create a value of DBInstance with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:engineVersion:DBInstance', dbInstance_engineVersion - Indicates the database engine version.

$sel:storageEncrypted:DBInstance', dbInstance_storageEncrypted - Specifies whether or not the instance is encrypted.

$sel:dbClusterIdentifier:DBInstance', dbInstance_dbClusterIdentifier - Contains the name of the cluster that the instance is a member of if the instance is a member of a cluster.

$sel:publiclyAccessible:DBInstance', dbInstance_publiclyAccessible - Not supported. Amazon DocumentDB does not currently support public endpoints. The value of PubliclyAccessible is always false.

$sel:autoMinorVersionUpgrade:DBInstance', dbInstance_autoMinorVersionUpgrade - Does not apply. This parameter does not apply to Amazon DocumentDB. Amazon DocumentDB does not perform minor version upgrades regardless of the value set.

$sel:dbInstanceArn:DBInstance', dbInstance_dbInstanceArn - The Amazon Resource Name (ARN) for the instance.

$sel:instanceCreateTime:DBInstance', dbInstance_instanceCreateTime - Provides the date and time that the instance was created.

$sel:engine:DBInstance', dbInstance_engine - Provides the name of the database engine to be used for this instance.

$sel:latestRestorableTime:DBInstance', dbInstance_latestRestorableTime - Specifies the latest time to which a database can be restored with point-in-time restore.

$sel:dbInstanceClass:DBInstance', dbInstance_dbInstanceClass - Contains the name of the compute and memory capacity class of the instance.

$sel:promotionTier:DBInstance', dbInstance_promotionTier - A value that specifies the order in which an Amazon DocumentDB replica is promoted to the primary instance after a failure of the existing primary instance.

$sel:preferredMaintenanceWindow:DBInstance', dbInstance_preferredMaintenanceWindow - Specifies the weekly time range during which system maintenance can occur, in Universal Coordinated Time (UTC).

$sel:cACertificateIdentifier:DBInstance', dbInstance_cACertificateIdentifier - The identifier of the CA certificate for this DB instance.

$sel:dbInstanceIdentifier:DBInstance', dbInstance_dbInstanceIdentifier - Contains a user-provided database identifier. This identifier is the unique key that identifies an instance.

$sel:kmsKeyId:DBInstance', dbInstance_kmsKeyId - If StorageEncrypted is true, the KMS key identifier for the encrypted instance.

$sel:preferredBackupWindow:DBInstance', dbInstance_preferredBackupWindow - Specifies the daily time range during which automated backups are created if automated backups are enabled, as determined by the BackupRetentionPeriod.

$sel:availabilityZone:DBInstance', dbInstance_availabilityZone - Specifies the name of the Availability Zone that the instance is located in.

$sel:vpcSecurityGroups:DBInstance', dbInstance_vpcSecurityGroups - Provides a list of VPC security group elements that the instance belongs to.

$sel:backupRetentionPeriod:DBInstance', dbInstance_backupRetentionPeriod - Specifies the number of days for which automatic snapshots are retained.

$sel:dbSubnetGroup:DBInstance', dbInstance_dbSubnetGroup - Specifies information on the subnet group that is associated with the instance, including the name, description, and subnets in the subnet group.

$sel:enabledCloudwatchLogsExports:DBInstance', dbInstance_enabledCloudwatchLogsExports - A list of log types that this instance is configured to export to CloudWatch Logs.

$sel:dbiResourceId:DBInstance', dbInstance_dbiResourceId - The Region-unique, immutable identifier for the instance. This identifier is found in CloudTrail log entries whenever the KMS key for the instance is accessed.

$sel:endpoint:DBInstance', dbInstance_endpoint - Specifies the connection endpoint.

$sel:dbInstanceStatus:DBInstance', dbInstance_dbInstanceStatus - Specifies the current state of this database.

$sel:pendingModifiedValues:DBInstance', dbInstance_pendingModifiedValues - Specifies that changes to the instance are pending. This element is included only when changes are pending. Specific changes are identified by subelements.

$sel:statusInfos:DBInstance', dbInstance_statusInfos - The status of a read replica. If the instance is not a read replica, this is blank.

dbInstance_engineVersion :: Lens' DBInstance (Maybe Text)

Indicates the database engine version.

dbInstance_storageEncrypted :: Lens' DBInstance (Maybe Bool)

Specifies whether or not the instance is encrypted.

dbInstance_dbClusterIdentifier :: Lens' DBInstance (Maybe Text)

Contains the name of the cluster that the instance is a member of if the instance is a member of a cluster.

dbInstance_publiclyAccessible :: Lens' DBInstance (Maybe Bool)

Not supported. Amazon DocumentDB does not currently support public endpoints. The value of PubliclyAccessible is always false.

dbInstance_autoMinorVersionUpgrade :: Lens' DBInstance (Maybe Bool)

Does not apply. This parameter does not apply to Amazon DocumentDB. Amazon DocumentDB does not perform minor version upgrades regardless of the value set.

dbInstance_dbInstanceArn :: Lens' DBInstance (Maybe Text)

The Amazon Resource Name (ARN) for the instance.

dbInstance_instanceCreateTime :: Lens' DBInstance (Maybe UTCTime)

Provides the date and time that the instance was created.

dbInstance_engine :: Lens' DBInstance (Maybe Text)

Provides the name of the database engine to be used for this instance.

dbInstance_latestRestorableTime :: Lens' DBInstance (Maybe UTCTime)

Specifies the latest time to which a database can be restored with point-in-time restore.

dbInstance_dbInstanceClass :: Lens' DBInstance (Maybe Text)

Contains the name of the compute and memory capacity class of the instance.

dbInstance_promotionTier :: Lens' DBInstance (Maybe Int)

A value that specifies the order in which an Amazon DocumentDB replica is promoted to the primary instance after a failure of the existing primary instance.

dbInstance_preferredMaintenanceWindow :: Lens' DBInstance (Maybe Text)

Specifies the weekly time range during which system maintenance can occur, in Universal Coordinated Time (UTC).

dbInstance_cACertificateIdentifier :: Lens' DBInstance (Maybe Text)

The identifier of the CA certificate for this DB instance.

dbInstance_dbInstanceIdentifier :: Lens' DBInstance (Maybe Text)

Contains a user-provided database identifier. This identifier is the unique key that identifies an instance.

dbInstance_kmsKeyId :: Lens' DBInstance (Maybe Text)

If StorageEncrypted is true, the KMS key identifier for the encrypted instance.

dbInstance_preferredBackupWindow :: Lens' DBInstance (Maybe Text)

Specifies the daily time range during which automated backups are created if automated backups are enabled, as determined by the BackupRetentionPeriod.

dbInstance_availabilityZone :: Lens' DBInstance (Maybe Text)

Specifies the name of the Availability Zone that the instance is located in.

dbInstance_vpcSecurityGroups :: Lens' DBInstance (Maybe [VpcSecurityGroupMembership])

Provides a list of VPC security group elements that the instance belongs to.

dbInstance_backupRetentionPeriod :: Lens' DBInstance (Maybe Int)

Specifies the number of days for which automatic snapshots are retained.

dbInstance_dbSubnetGroup :: Lens' DBInstance (Maybe DBSubnetGroup)

Specifies information on the subnet group that is associated with the instance, including the name, description, and subnets in the subnet group.

dbInstance_enabledCloudwatchLogsExports :: Lens' DBInstance (Maybe [Text])

A list of log types that this instance is configured to export to CloudWatch Logs.

dbInstance_dbiResourceId :: Lens' DBInstance (Maybe Text)

The Region-unique, immutable identifier for the instance. This identifier is found in CloudTrail log entries whenever the KMS key for the instance is accessed.

dbInstance_endpoint :: Lens' DBInstance (Maybe Endpoint)

Specifies the connection endpoint.

dbInstance_dbInstanceStatus :: Lens' DBInstance (Maybe Text)

Specifies the current state of this database.

dbInstance_pendingModifiedValues :: Lens' DBInstance (Maybe PendingModifiedValues)

Specifies that changes to the instance are pending. This element is included only when changes are pending. Specific changes are identified by subelements.

dbInstance_statusInfos :: Lens' DBInstance (Maybe [DBInstanceStatusInfo])

The status of a read replica. If the instance is not a read replica, this is blank.