Amazonka.DocumentDB.Types.Event

Description

Documentation

data Event

Detailed information about an event.

See: newEvent smart constructor.

Constructors

Event'

Fields

sourceType :: Maybe SourceType

Specifies the source type for this event.

sourceArn :: Maybe Text

The Amazon Resource Name (ARN) for the event.

sourceIdentifier :: Maybe Text

Provides the identifier for the source of the event.

date :: Maybe ISO8601

Specifies the date and time of the event.

eventCategories :: Maybe [Text]

Specifies the category for the event.

message :: Maybe Text

Provides the text of this event.

newEvent :: Event

Create a value of Event with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:sourceType:Event', event_sourceType - Specifies the source type for this event.

$sel:sourceArn:Event', event_sourceArn - The Amazon Resource Name (ARN) for the event.

$sel:sourceIdentifier:Event', event_sourceIdentifier - Provides the identifier for the source of the event.

$sel:date:Event', event_date - Specifies the date and time of the event.

$sel:eventCategories:Event', event_eventCategories - Specifies the category for the event.

$sel:message:Event', event_message - Provides the text of this event.

event_sourceType :: Lens' Event (Maybe SourceType)

Specifies the source type for this event.

event_sourceArn :: Lens' Event (Maybe Text)

The Amazon Resource Name (ARN) for the event.

event_sourceIdentifier :: Lens' Event (Maybe Text)

Provides the identifier for the source of the event.

event_date :: Lens' Event (Maybe UTCTime)

Specifies the date and time of the event.

event_eventCategories :: Lens' Event (Maybe [Text])

Specifies the category for the event.

event_message :: Lens' Event (Maybe Text)

Provides the text of this event.