Amazonka.DocumentDB.Types.GlobalCluster

Description

Documentation

data GlobalCluster

A data type representing an Amazon DocumentDB global cluster.

See: newGlobalCluster smart constructor.

Constructors

GlobalCluster'

Fields

engineVersion :: Maybe Text

Indicates the database engine version.

status :: Maybe Text

Specifies the current state of this global cluster.

deletionProtection :: Maybe Bool

The deletion protection setting for the new global cluster.

storageEncrypted :: Maybe Bool

The storage encryption setting for the global cluster.

globalClusterIdentifier :: Maybe Text

Contains a user-supplied global cluster identifier. This identifier is the unique key that identifies a global cluster.

engine :: Maybe Text

The Amazon DocumentDB database engine used by the global cluster.

globalClusterArn :: Maybe Text

The Amazon Resource Name (ARN) for the global cluster.

databaseName :: Maybe Text

The default database name within the new global cluster.

globalClusterMembers :: Maybe [GlobalClusterMember]

The list of cluster IDs for secondary clusters within the global cluster. Currently limited to one item.

globalClusterResourceId :: Maybe Text

The Region-unique, immutable identifier for the global database cluster. This identifier is found in AWS CloudTrail log entries whenever the AWS KMS customer master key (CMK) for the cluster is accessed.

newGlobalCluster :: GlobalCluster

Create a value of GlobalCluster with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:engineVersion:GlobalCluster', globalCluster_engineVersion - Indicates the database engine version.

$sel:status:GlobalCluster', globalCluster_status - Specifies the current state of this global cluster.

$sel:deletionProtection:GlobalCluster', globalCluster_deletionProtection - The deletion protection setting for the new global cluster.

$sel:storageEncrypted:GlobalCluster', globalCluster_storageEncrypted - The storage encryption setting for the global cluster.

$sel:globalClusterIdentifier:GlobalCluster', globalCluster_globalClusterIdentifier - Contains a user-supplied global cluster identifier. This identifier is the unique key that identifies a global cluster.

$sel:engine:GlobalCluster', globalCluster_engine - The Amazon DocumentDB database engine used by the global cluster.

$sel:globalClusterArn:GlobalCluster', globalCluster_globalClusterArn - The Amazon Resource Name (ARN) for the global cluster.

$sel:databaseName:GlobalCluster', globalCluster_databaseName - The default database name within the new global cluster.

$sel:globalClusterMembers:GlobalCluster', globalCluster_globalClusterMembers - The list of cluster IDs for secondary clusters within the global cluster. Currently limited to one item.

$sel:globalClusterResourceId:GlobalCluster', globalCluster_globalClusterResourceId - The Region-unique, immutable identifier for the global database cluster. This identifier is found in AWS CloudTrail log entries whenever the AWS KMS customer master key (CMK) for the cluster is accessed.

globalCluster_engineVersion :: Lens' GlobalCluster (Maybe Text)

Indicates the database engine version.

globalCluster_status :: Lens' GlobalCluster (Maybe Text)

Specifies the current state of this global cluster.

globalCluster_deletionProtection :: Lens' GlobalCluster (Maybe Bool)

The deletion protection setting for the new global cluster.

globalCluster_storageEncrypted :: Lens' GlobalCluster (Maybe Bool)

The storage encryption setting for the global cluster.

globalCluster_globalClusterIdentifier :: Lens' GlobalCluster (Maybe Text)

Contains a user-supplied global cluster identifier. This identifier is the unique key that identifies a global cluster.

globalCluster_engine :: Lens' GlobalCluster (Maybe Text)

The Amazon DocumentDB database engine used by the global cluster.

globalCluster_globalClusterArn :: Lens' GlobalCluster (Maybe Text)

The Amazon Resource Name (ARN) for the global cluster.

globalCluster_databaseName :: Lens' GlobalCluster (Maybe Text)

The default database name within the new global cluster.

globalCluster_globalClusterMembers :: Lens' GlobalCluster (Maybe [GlobalClusterMember])

The list of cluster IDs for secondary clusters within the global cluster. Currently limited to one item.

globalCluster_globalClusterResourceId :: Lens' GlobalCluster (Maybe Text)

The Region-unique, immutable identifier for the global database cluster. This identifier is found in AWS CloudTrail log entries whenever the AWS KMS customer master key (CMK) for the cluster is accessed.