Amazonka.DocumentDB.Types.OrderableDBInstanceOption

Description

Documentation

data OrderableDBInstanceOption

The options that are available for an instance.

See: newOrderableDBInstanceOption smart constructor.

Constructors

OrderableDBInstanceOption'

Fields

engineVersion :: Maybe Text

The engine version of an instance.

engine :: Maybe Text

The engine type of an instance.

dbInstanceClass :: Maybe Text

The instance class for an instance.

licenseModel :: Maybe Text

The license model for an instance.

availabilityZones :: Maybe [AvailabilityZone]

A list of Availability Zones for an instance.

vpc :: Maybe Bool

Indicates whether an instance is in a virtual private cloud (VPC).

newOrderableDBInstanceOption :: OrderableDBInstanceOption

Create a value of OrderableDBInstanceOption with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:engineVersion:OrderableDBInstanceOption', orderableDBInstanceOption_engineVersion - The engine version of an instance.

$sel:engine:OrderableDBInstanceOption', orderableDBInstanceOption_engine - The engine type of an instance.

$sel:dbInstanceClass:OrderableDBInstanceOption', orderableDBInstanceOption_dbInstanceClass - The instance class for an instance.

$sel:licenseModel:OrderableDBInstanceOption', orderableDBInstanceOption_licenseModel - The license model for an instance.

$sel:availabilityZones:OrderableDBInstanceOption', orderableDBInstanceOption_availabilityZones - A list of Availability Zones for an instance.

$sel:vpc:OrderableDBInstanceOption', orderableDBInstanceOption_vpc - Indicates whether an instance is in a virtual private cloud (VPC).

orderableDBInstanceOption_engineVersion :: Lens' OrderableDBInstanceOption (Maybe Text)

The engine version of an instance.

orderableDBInstanceOption_engine :: Lens' OrderableDBInstanceOption (Maybe Text)

The engine type of an instance.

orderableDBInstanceOption_dbInstanceClass :: Lens' OrderableDBInstanceOption (Maybe Text)

The instance class for an instance.

orderableDBInstanceOption_licenseModel :: Lens' OrderableDBInstanceOption (Maybe Text)

The license model for an instance.

orderableDBInstanceOption_availabilityZones :: Lens' OrderableDBInstanceOption (Maybe [AvailabilityZone])

A list of Availability Zones for an instance.

orderableDBInstanceOption_vpc :: Lens' OrderableDBInstanceOption (Maybe Bool)

Indicates whether an instance is in a virtual private cloud (VPC).