Amazonka.DocumentDB.Types.Parameter

Description

Documentation

data Parameter

Detailed information about an individual parameter.

See: newParameter smart constructor.

Constructors

Parameter'

Fields

applyType :: Maybe Text

Specifies the engine-specific parameters type.

parameterValue :: Maybe Text

Specifies the value of the parameter.

applyMethod :: Maybe ApplyMethod

Indicates when to apply parameter updates.

minimumEngineVersion :: Maybe Text

The earliest engine version to which the parameter can apply.

source :: Maybe Text

Indicates the source of the parameter value.

isModifiable :: Maybe Bool

Indicates whether (true) or not (false) the parameter can be modified. Some parameters have security or operational implications that prevent them from being changed.

dataType :: Maybe Text

Specifies the valid data type for the parameter.

allowedValues :: Maybe Text

Specifies the valid range of values for the parameter.

parameterName :: Maybe Text

Specifies the name of the parameter.

description :: Maybe Text

Provides a description of the parameter.

newParameter :: Parameter

Create a value of Parameter with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:applyType:Parameter', parameter_applyType - Specifies the engine-specific parameters type.

$sel:parameterValue:Parameter', parameter_parameterValue - Specifies the value of the parameter.

$sel:applyMethod:Parameter', parameter_applyMethod - Indicates when to apply parameter updates.

$sel:minimumEngineVersion:Parameter', parameter_minimumEngineVersion - The earliest engine version to which the parameter can apply.

$sel:source:Parameter', parameter_source - Indicates the source of the parameter value.

$sel:isModifiable:Parameter', parameter_isModifiable - Indicates whether (true) or not (false) the parameter can be modified. Some parameters have security or operational implications that prevent them from being changed.

$sel:dataType:Parameter', parameter_dataType - Specifies the valid data type for the parameter.

$sel:allowedValues:Parameter', parameter_allowedValues - Specifies the valid range of values for the parameter.

$sel:parameterName:Parameter', parameter_parameterName - Specifies the name of the parameter.

$sel:description:Parameter', parameter_description - Provides a description of the parameter.

parameter_applyType :: Lens' Parameter (Maybe Text)

Specifies the engine-specific parameters type.

parameter_parameterValue :: Lens' Parameter (Maybe Text)

Specifies the value of the parameter.

parameter_applyMethod :: Lens' Parameter (Maybe ApplyMethod)

Indicates when to apply parameter updates.

parameter_minimumEngineVersion :: Lens' Parameter (Maybe Text)

The earliest engine version to which the parameter can apply.

parameter_source :: Lens' Parameter (Maybe Text)

Indicates the source of the parameter value.

parameter_isModifiable :: Lens' Parameter (Maybe Bool)

Indicates whether (true) or not (false) the parameter can be modified. Some parameters have security or operational implications that prevent them from being changed.

parameter_dataType :: Lens' Parameter (Maybe Text)

Specifies the valid data type for the parameter.

parameter_allowedValues :: Lens' Parameter (Maybe Text)

Specifies the valid range of values for the parameter.

parameter_parameterName :: Lens' Parameter (Maybe Text)

Specifies the name of the parameter.

parameter_description :: Lens' Parameter (Maybe Text)

Provides a description of the parameter.