Amazonka.DynamoDB.CreateTable

Description

The CreateTable operation adds a new table to your account. In an AWS account, table names must be unique within each Region. That is, you can have two tables with same name if you create the tables in different Regions.

CreateTable is an asynchronous operation. Upon receiving a CreateTable request, DynamoDB immediately returns a response with a TableStatus of CREATING. After the table is created, DynamoDB sets the TableStatus to ACTIVE. You can perform read and write operations only on an ACTIVE table.

You can optionally define secondary indexes on the new table, as part of the CreateTable operation. If you want to create multiple tables with secondary indexes on them, you must create the tables sequentially. Only one table with secondary indexes can be in the CREATING state at any given time.

You can use the DescribeTable action to check the table status.

Creating a Request

data CreateTable

Represents the input of a CreateTable operation.

See: newCreateTable smart constructor.

Constructors

CreateTable'

Fields

provisionedThroughput :: Maybe ProvisionedThroughput

Represents the provisioned throughput settings for a specified table or index. The settings can be modified using the UpdateTable operation.

If you set BillingMode as PROVISIONED, you must specify this property. If you set BillingMode as PAY_PER_REQUEST, you cannot specify this property.

For current minimum and maximum provisioned throughput values, see Service, Account, and Table Quotas in the Amazon DynamoDB Developer Guide.

sSESpecification :: Maybe SSESpecification

Represents the settings used to enable server-side encryption.

globalSecondaryIndexes :: Maybe [GlobalSecondaryIndex]

One or more global secondary indexes (the maximum is 20) to be created on the table. Each global secondary index in the array includes the following:

IndexName - The name of the global secondary index. Must be unique only for this table.

KeySchema - Specifies the key schema for the global secondary index.

Projection - Specifies attributes that are copied (projected) from the table into the index. These are in addition to the primary key attributes and index key attributes, which are automatically projected. Each attribute specification is composed of:

ProjectionType - One of the following:

KEYS_ONLY - Only the index and primary keys are projected into the index.

INCLUDE - Only the specified table attributes are projected into the index. The list of projected attributes is in NonKeyAttributes.

ALL - All of the table attributes are projected into the index.

NonKeyAttributes - A list of one or more non-key attribute names that are projected into the secondary index. The total count of attributes provided in NonKeyAttributes, summed across all of the secondary indexes, must not exceed 100. If you project the same attribute into two different indexes, this counts as two distinct attributes when determining the total.

ProvisionedThroughput - The provisioned throughput settings for the global secondary index, consisting of read and write capacity units.

localSecondaryIndexes :: Maybe [LocalSecondaryIndex]

One or more local secondary indexes (the maximum is 5) to be created on the table. Each index is scoped to a given partition key value. There is a 10 GB size limit per partition key value; otherwise, the size of a local secondary index is unconstrained.

Each local secondary index in the array includes the following:

IndexName - The name of the local secondary index. Must be unique only for this table.

KeySchema - Specifies the key schema for the local secondary index. The key schema must begin with the same partition key as the table.

Projection - Specifies attributes that are copied (projected) from the table into the index. These are in addition to the primary key attributes and index key attributes, which are automatically projected. Each attribute specification is composed of:

ProjectionType - One of the following:

KEYS_ONLY - Only the index and primary keys are projected into the index.

INCLUDE - Only the specified table attributes are projected into the index. The list of projected attributes is in NonKeyAttributes.

ALL - All of the table attributes are projected into the index.

NonKeyAttributes - A list of one or more non-key attribute names that are projected into the secondary index. The total count of attributes provided in NonKeyAttributes, summed across all of the secondary indexes, must not exceed 100. If you project the same attribute into two different indexes, this counts as two distinct attributes when determining the total.

billingMode :: Maybe BillingMode

Controls how you are charged for read and write throughput and how you manage capacity. This setting can be changed later.

PROVISIONED - We recommend using PROVISIONED for predictable workloads. PROVISIONED sets the billing mode to Provisioned Mode.

PAY_PER_REQUEST - We recommend using PAY_PER_REQUEST for unpredictable workloads. PAY_PER_REQUEST sets the billing mode to On-Demand Mode.

tags :: Maybe [Tag]

A list of key-value pairs to label the table. For more information, see Tagging for DynamoDB.

streamSpecification :: Maybe StreamSpecification

The settings for DynamoDB Streams on the table. These settings consist of:

StreamEnabled - Indicates whether DynamoDB Streams is to be enabled (true) or disabled (false).

StreamViewType - When an item in the table is modified, StreamViewType determines what information is written to the table's stream. Valid values for StreamViewType are:

KEYS_ONLY - Only the key attributes of the modified item are written to the stream.

NEW_IMAGE - The entire item, as it appears after it was modified, is written to the stream.

OLD_IMAGE - The entire item, as it appeared before it was modified, is written to the stream.

NEW_AND_OLD_IMAGES - Both the new and the old item images of the item are written to the stream.

attributeDefinitions :: [AttributeDefinition]

An array of attributes that describe the key schema for the table and indexes.

tableName :: Text

The name of the table to create.

keySchema :: NonEmpty KeySchemaElement

Specifies the attributes that make up the primary key for a table or an index. The attributes in KeySchema must also be defined in the AttributeDefinitions array. For more information, see Data Model in the Amazon DynamoDB Developer Guide.

Each KeySchemaElement in the array is composed of:

AttributeName - The name of this key attribute.

KeyType - The role that the key attribute will assume:

HASH - partition key

RANGE - sort key

The partition key of an item is also known as its hash attribute. The term "hash attribute" derives from the DynamoDB usage of an internal hash function to evenly distribute data items across partitions, based on their partition key values.

The sort key of an item is also known as its range attribute. The term "range attribute" derives from the way DynamoDB stores items with the same partition key physically close together, in sorted order by the sort key value.

For a simple primary key (partition key), you must provide exactly one element with a KeyType of HASH.

For a composite primary key (partition key and sort key), you must provide exactly two elements, in this order: The first element must have a KeyType of HASH, and the second element must have a KeyType of RANGE.

For more information, see Working with Tables in the Amazon DynamoDB Developer Guide.

newCreateTable

Arguments

:: Text
$sel:tableName:CreateTable'

-> NonEmpty KeySchemaElement
$sel:keySchema:CreateTable'

-> CreateTable

Create a value of CreateTable with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:provisionedThroughput:CreateTable', createTable_provisionedThroughput - Represents the provisioned throughput settings for a specified table or index. The settings can be modified using the UpdateTable operation.

If you set BillingMode as PROVISIONED, you must specify this property. If you set BillingMode as PAY_PER_REQUEST, you cannot specify this property.

For current minimum and maximum provisioned throughput values, see Service, Account, and Table Quotas in the Amazon DynamoDB Developer Guide.

$sel:sSESpecification:CreateTable', createTable_sSESpecification - Represents the settings used to enable server-side encryption.

$sel:globalSecondaryIndexes:CreateTable', createTable_globalSecondaryIndexes - One or more global secondary indexes (the maximum is 20) to be created on the table. Each global secondary index in the array includes the following:

IndexName - The name of the global secondary index. Must be unique only for this table.

KeySchema - Specifies the key schema for the global secondary index.

Projection - Specifies attributes that are copied (projected) from the table into the index. These are in addition to the primary key attributes and index key attributes, which are automatically projected. Each attribute specification is composed of:

ProjectionType - One of the following:

KEYS_ONLY - Only the index and primary keys are projected into the index.

INCLUDE - Only the specified table attributes are projected into the index. The list of projected attributes is in NonKeyAttributes.

ALL - All of the table attributes are projected into the index.

NonKeyAttributes - A list of one or more non-key attribute names that are projected into the secondary index. The total count of attributes provided in NonKeyAttributes, summed across all of the secondary indexes, must not exceed 100. If you project the same attribute into two different indexes, this counts as two distinct attributes when determining the total.

ProvisionedThroughput - The provisioned throughput settings for the global secondary index, consisting of read and write capacity units.

$sel:localSecondaryIndexes:CreateTable', createTable_localSecondaryIndexes - One or more local secondary indexes (the maximum is 5) to be created on the table. Each index is scoped to a given partition key value. There is a 10 GB size limit per partition key value; otherwise, the size of a local secondary index is unconstrained.

Each local secondary index in the array includes the following:

IndexName - The name of the local secondary index. Must be unique only for this table.

KeySchema - Specifies the key schema for the local secondary index. The key schema must begin with the same partition key as the table.

Projection - Specifies attributes that are copied (projected) from the table into the index. These are in addition to the primary key attributes and index key attributes, which are automatically projected. Each attribute specification is composed of:

ProjectionType - One of the following:

KEYS_ONLY - Only the index and primary keys are projected into the index.

INCLUDE - Only the specified table attributes are projected into the index. The list of projected attributes is in NonKeyAttributes.

ALL - All of the table attributes are projected into the index.

NonKeyAttributes - A list of one or more non-key attribute names that are projected into the secondary index. The total count of attributes provided in NonKeyAttributes, summed across all of the secondary indexes, must not exceed 100. If you project the same attribute into two different indexes, this counts as two distinct attributes when determining the total.

$sel:billingMode:CreateTable', createTable_billingMode - Controls how you are charged for read and write throughput and how you manage capacity. This setting can be changed later.

PROVISIONED - We recommend using PROVISIONED for predictable workloads. PROVISIONED sets the billing mode to Provisioned Mode.

PAY_PER_REQUEST - We recommend using PAY_PER_REQUEST for unpredictable workloads. PAY_PER_REQUEST sets the billing mode to On-Demand Mode.

$sel:tags:CreateTable', createTable_tags - A list of key-value pairs to label the table. For more information, see Tagging for DynamoDB.

$sel:streamSpecification:CreateTable', createTable_streamSpecification - The settings for DynamoDB Streams on the table. These settings consist of:

StreamEnabled - Indicates whether DynamoDB Streams is to be enabled (true) or disabled (false).

StreamViewType - When an item in the table is modified, StreamViewType determines what information is written to the table's stream. Valid values for StreamViewType are:

KEYS_ONLY - Only the key attributes of the modified item are written to the stream.

NEW_IMAGE - The entire item, as it appears after it was modified, is written to the stream.

OLD_IMAGE - The entire item, as it appeared before it was modified, is written to the stream.

NEW_AND_OLD_IMAGES - Both the new and the old item images of the item are written to the stream.

$sel:attributeDefinitions:CreateTable', createTable_attributeDefinitions - An array of attributes that describe the key schema for the table and indexes.

$sel:tableName:CreateTable', createTable_tableName - The name of the table to create.

$sel:keySchema:CreateTable', createTable_keySchema - Specifies the attributes that make up the primary key for a table or an index. The attributes in KeySchema must also be defined in the AttributeDefinitions array. For more information, see Data Model in the Amazon DynamoDB Developer Guide.

Each KeySchemaElement in the array is composed of:

AttributeName - The name of this key attribute.

KeyType - The role that the key attribute will assume:

HASH - partition key

RANGE - sort key

The partition key of an item is also known as its hash attribute. The term "hash attribute" derives from the DynamoDB usage of an internal hash function to evenly distribute data items across partitions, based on their partition key values.

The sort key of an item is also known as its range attribute. The term "range attribute" derives from the way DynamoDB stores items with the same partition key physically close together, in sorted order by the sort key value.

For a simple primary key (partition key), you must provide exactly one element with a KeyType of HASH.

For a composite primary key (partition key and sort key), you must provide exactly two elements, in this order: The first element must have a KeyType of HASH, and the second element must have a KeyType of RANGE.

For more information, see Working with Tables in the Amazon DynamoDB Developer Guide.

Request Lenses

createTable_provisionedThroughput :: Lens' CreateTable (Maybe ProvisionedThroughput)

Represents the provisioned throughput settings for a specified table or index. The settings can be modified using the UpdateTable operation.

If you set BillingMode as PROVISIONED, you must specify this property. If you set BillingMode as PAY_PER_REQUEST, you cannot specify this property.

For current minimum and maximum provisioned throughput values, see Service, Account, and Table Quotas in the Amazon DynamoDB Developer Guide.

createTable_sSESpecification :: Lens' CreateTable (Maybe SSESpecification)

Represents the settings used to enable server-side encryption.

createTable_globalSecondaryIndexes :: Lens' CreateTable (Maybe [GlobalSecondaryIndex])

One or more global secondary indexes (the maximum is 20) to be created on the table. Each global secondary index in the array includes the following:

IndexName - The name of the global secondary index. Must be unique only for this table.

KeySchema - Specifies the key schema for the global secondary index.

Projection - Specifies attributes that are copied (projected) from the table into the index. These are in addition to the primary key attributes and index key attributes, which are automatically projected. Each attribute specification is composed of:

ProjectionType - One of the following:

KEYS_ONLY - Only the index and primary keys are projected into the index.

INCLUDE - Only the specified table attributes are projected into the index. The list of projected attributes is in NonKeyAttributes.

ALL - All of the table attributes are projected into the index.

NonKeyAttributes - A list of one or more non-key attribute names that are projected into the secondary index. The total count of attributes provided in NonKeyAttributes, summed across all of the secondary indexes, must not exceed 100. If you project the same attribute into two different indexes, this counts as two distinct attributes when determining the total.

ProvisionedThroughput - The provisioned throughput settings for the global secondary index, consisting of read and write capacity units.

createTable_localSecondaryIndexes :: Lens' CreateTable (Maybe [LocalSecondaryIndex])

One or more local secondary indexes (the maximum is 5) to be created on the table. Each index is scoped to a given partition key value. There is a 10 GB size limit per partition key value; otherwise, the size of a local secondary index is unconstrained.

Each local secondary index in the array includes the following:

IndexName - The name of the local secondary index. Must be unique only for this table.

KeySchema - Specifies the key schema for the local secondary index. The key schema must begin with the same partition key as the table.

Projection - Specifies attributes that are copied (projected) from the table into the index. These are in addition to the primary key attributes and index key attributes, which are automatically projected. Each attribute specification is composed of:

ProjectionType - One of the following:

KEYS_ONLY - Only the index and primary keys are projected into the index.

INCLUDE - Only the specified table attributes are projected into the index. The list of projected attributes is in NonKeyAttributes.

ALL - All of the table attributes are projected into the index.

NonKeyAttributes - A list of one or more non-key attribute names that are projected into the secondary index. The total count of attributes provided in NonKeyAttributes, summed across all of the secondary indexes, must not exceed 100. If you project the same attribute into two different indexes, this counts as two distinct attributes when determining the total.

createTable_billingMode :: Lens' CreateTable (Maybe BillingMode)

Controls how you are charged for read and write throughput and how you manage capacity. This setting can be changed later.

PROVISIONED - We recommend using PROVISIONED for predictable workloads. PROVISIONED sets the billing mode to Provisioned Mode.

PAY_PER_REQUEST - We recommend using PAY_PER_REQUEST for unpredictable workloads. PAY_PER_REQUEST sets the billing mode to On-Demand Mode.

createTable_tags :: Lens' CreateTable (Maybe [Tag])

A list of key-value pairs to label the table. For more information, see Tagging for DynamoDB.

createTable_streamSpecification :: Lens' CreateTable (Maybe StreamSpecification)

The settings for DynamoDB Streams on the table. These settings consist of:

StreamEnabled - Indicates whether DynamoDB Streams is to be enabled (true) or disabled (false).

StreamViewType - When an item in the table is modified, StreamViewType determines what information is written to the table's stream. Valid values for StreamViewType are:

KEYS_ONLY - Only the key attributes of the modified item are written to the stream.

NEW_IMAGE - The entire item, as it appears after it was modified, is written to the stream.

OLD_IMAGE - The entire item, as it appeared before it was modified, is written to the stream.

NEW_AND_OLD_IMAGES - Both the new and the old item images of the item are written to the stream.

createTable_attributeDefinitions :: Lens' CreateTable [AttributeDefinition]

An array of attributes that describe the key schema for the table and indexes.

createTable_tableName :: Lens' CreateTable Text

The name of the table to create.

createTable_keySchema :: Lens' CreateTable (NonEmpty KeySchemaElement)

Specifies the attributes that make up the primary key for a table or an index. The attributes in KeySchema must also be defined in the AttributeDefinitions array. For more information, see Data Model in the Amazon DynamoDB Developer Guide.

Each KeySchemaElement in the array is composed of:

AttributeName - The name of this key attribute.

KeyType - The role that the key attribute will assume:

HASH - partition key

RANGE - sort key

The partition key of an item is also known as its hash attribute. The term "hash attribute" derives from the DynamoDB usage of an internal hash function to evenly distribute data items across partitions, based on their partition key values.

The sort key of an item is also known as its range attribute. The term "range attribute" derives from the way DynamoDB stores items with the same partition key physically close together, in sorted order by the sort key value.

For a simple primary key (partition key), you must provide exactly one element with a KeyType of HASH.

For a composite primary key (partition key and sort key), you must provide exactly two elements, in this order: The first element must have a KeyType of HASH, and the second element must have a KeyType of RANGE.

For more information, see Working with Tables in the Amazon DynamoDB Developer Guide.

Destructuring the Response

data CreateTableResponse

Represents the output of a CreateTable operation.

See: newCreateTableResponse smart constructor.

Constructors

CreateTableResponse'

Fields

tableDescription :: Maybe TableDescription

Represents the properties of the table.

httpStatus :: Int

The response's http status code.

newCreateTableResponse

Arguments

:: Int
$sel:httpStatus:CreateTableResponse'

-> CreateTableResponse

Create a value of CreateTableResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:tableDescription:CreateTableResponse', createTableResponse_tableDescription - Represents the properties of the table.

$sel:httpStatus:CreateTableResponse', createTableResponse_httpStatus - The response's http status code.

Response Lenses

createTableResponse_tableDescription :: Lens' CreateTableResponse (Maybe TableDescription)

Represents the properties of the table.

createTableResponse_httpStatus :: Lens' CreateTableResponse Int

The response's http status code.