Amazonka.DynamoDB.ExecuteTransaction

Description

This operation allows you to perform transactional reads or writes on data stored in DynamoDB, using PartiQL.

Creating a Request

data ExecuteTransaction

See: newExecuteTransaction smart constructor.

Constructors

ExecuteTransaction'

Fields

clientRequestToken :: Maybe Text

Set this value to get remaining results, if NextToken was returned in the statement response.

transactStatements :: NonEmpty ParameterizedStatement

The list of PartiQL statements representing the transaction to run.

newExecuteTransaction

Arguments

:: NonEmpty ParameterizedStatement
$sel:transactStatements:ExecuteTransaction'

-> ExecuteTransaction

Create a value of ExecuteTransaction with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:clientRequestToken:ExecuteTransaction', executeTransaction_clientRequestToken - Set this value to get remaining results, if NextToken was returned in the statement response.

$sel:transactStatements:ExecuteTransaction', executeTransaction_transactStatements - The list of PartiQL statements representing the transaction to run.

Request Lenses

executeTransaction_clientRequestToken :: Lens' ExecuteTransaction (Maybe Text)

Set this value to get remaining results, if NextToken was returned in the statement response.

executeTransaction_transactStatements :: Lens' ExecuteTransaction (NonEmpty ParameterizedStatement)

The list of PartiQL statements representing the transaction to run.

Destructuring the Response

data ExecuteTransactionResponse

See: newExecuteTransactionResponse smart constructor.

Constructors

ExecuteTransactionResponse'

Fields

responses :: Maybe (NonEmpty ItemResponse)

The response to a PartiQL transaction.

httpStatus :: Int

The response's http status code.

newExecuteTransactionResponse

Arguments

:: Int
$sel:httpStatus:ExecuteTransactionResponse'

-> ExecuteTransactionResponse

Create a value of ExecuteTransactionResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:responses:ExecuteTransactionResponse', executeTransactionResponse_responses - The response to a PartiQL transaction.

$sel:httpStatus:ExecuteTransactionResponse', executeTransactionResponse_httpStatus - The response's http status code.

Response Lenses

executeTransactionResponse_responses :: Lens' ExecuteTransactionResponse (Maybe (NonEmpty ItemResponse))

The response to a PartiQL transaction.

executeTransactionResponse_httpStatus :: Lens' ExecuteTransactionResponse Int

The response's http status code.