Amazonka.DynamoDB.ExportTableToPointInTime

Description

Exports table data to an S3 bucket. The table must have point in time recovery enabled, and you can export data from any time within the point in time recovery window.

Creating a Request

data ExportTableToPointInTime

See: newExportTableToPointInTime smart constructor.

Constructors

ExportTableToPointInTime'

Fields

s3BucketOwner :: Maybe Text

The ID of the AWS account that owns the bucket the export will be stored in.

exportFormat :: Maybe ExportFormat

The format for the exported data. Valid values for ExportFormat are DYNAMODB_JSON or ION.

s3SseKmsKeyId :: Maybe Text

The ID of the AWS KMS managed key used to encrypt the S3 bucket where export data will be stored (if applicable).

clientToken :: Maybe Text

Providing a ClientToken makes the call to ExportTableToPointInTimeInput idempotent, meaning that multiple identical calls have the same effect as one single call.

A client token is valid for 8 hours after the first request that uses it is completed. After 8 hours, any request with the same client token is treated as a new request. Do not resubmit the same request with the same client token for more than 8 hours, or the result might not be idempotent.

If you submit a request with the same client token but a change in other parameters within the 8-hour idempotency window, DynamoDB returns an IdempotentParameterMismatch exception.

exportTime :: Maybe POSIX

Time in the past from which to export table data. The table export will be a snapshot of the table's state at this point in time.

s3SseAlgorithm :: Maybe S3SseAlgorithm

Type of encryption used on the bucket where export data will be stored. Valid values for S3SseAlgorithm are:

AES256 - server-side encryption with Amazon S3 managed keys

KMS - server-side encryption with AWS KMS managed keys

s3Prefix :: Maybe Text

The Amazon S3 bucket prefix to use as the file name and path of the exported snapshot.

tableArn :: Text

The Amazon Resource Name (ARN) associated with the table to export.

s3Bucket :: Text

The name of the Amazon S3 bucket to export the snapshot to.

newExportTableToPointInTime

Arguments

:: Text
$sel:tableArn:ExportTableToPointInTime'

-> Text
$sel:s3Bucket:ExportTableToPointInTime'

-> ExportTableToPointInTime

Create a value of ExportTableToPointInTime with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:s3BucketOwner:ExportTableToPointInTime', exportTableToPointInTime_s3BucketOwner - The ID of the AWS account that owns the bucket the export will be stored in.

$sel:exportFormat:ExportTableToPointInTime', exportTableToPointInTime_exportFormat - The format for the exported data. Valid values for ExportFormat are DYNAMODB_JSON or ION.

$sel:s3SseKmsKeyId:ExportTableToPointInTime', exportTableToPointInTime_s3SseKmsKeyId - The ID of the AWS KMS managed key used to encrypt the S3 bucket where export data will be stored (if applicable).

$sel:clientToken:ExportTableToPointInTime', exportTableToPointInTime_clientToken - Providing a ClientToken makes the call to ExportTableToPointInTimeInput idempotent, meaning that multiple identical calls have the same effect as one single call.

A client token is valid for 8 hours after the first request that uses it is completed. After 8 hours, any request with the same client token is treated as a new request. Do not resubmit the same request with the same client token for more than 8 hours, or the result might not be idempotent.

If you submit a request with the same client token but a change in other parameters within the 8-hour idempotency window, DynamoDB returns an IdempotentParameterMismatch exception.

$sel:exportTime:ExportTableToPointInTime', exportTableToPointInTime_exportTime - Time in the past from which to export table data. The table export will be a snapshot of the table's state at this point in time.

$sel:s3SseAlgorithm:ExportTableToPointInTime', exportTableToPointInTime_s3SseAlgorithm - Type of encryption used on the bucket where export data will be stored. Valid values for S3SseAlgorithm are:

AES256 - server-side encryption with Amazon S3 managed keys

KMS - server-side encryption with AWS KMS managed keys

$sel:s3Prefix:ExportTableToPointInTime', exportTableToPointInTime_s3Prefix - The Amazon S3 bucket prefix to use as the file name and path of the exported snapshot.

$sel:tableArn:ExportTableToPointInTime', exportTableToPointInTime_tableArn - The Amazon Resource Name (ARN) associated with the table to export.

$sel:s3Bucket:ExportTableToPointInTime', exportTableToPointInTime_s3Bucket - The name of the Amazon S3 bucket to export the snapshot to.

Request Lenses

exportTableToPointInTime_s3BucketOwner :: Lens' ExportTableToPointInTime (Maybe Text)

The ID of the AWS account that owns the bucket the export will be stored in.

exportTableToPointInTime_exportFormat :: Lens' ExportTableToPointInTime (Maybe ExportFormat)

The format for the exported data. Valid values for ExportFormat are DYNAMODB_JSON or ION.

exportTableToPointInTime_s3SseKmsKeyId :: Lens' ExportTableToPointInTime (Maybe Text)

The ID of the AWS KMS managed key used to encrypt the S3 bucket where export data will be stored (if applicable).

exportTableToPointInTime_clientToken :: Lens' ExportTableToPointInTime (Maybe Text)

Providing a ClientToken makes the call to ExportTableToPointInTimeInput idempotent, meaning that multiple identical calls have the same effect as one single call.

A client token is valid for 8 hours after the first request that uses it is completed. After 8 hours, any request with the same client token is treated as a new request. Do not resubmit the same request with the same client token for more than 8 hours, or the result might not be idempotent.

If you submit a request with the same client token but a change in other parameters within the 8-hour idempotency window, DynamoDB returns an IdempotentParameterMismatch exception.

exportTableToPointInTime_exportTime :: Lens' ExportTableToPointInTime (Maybe UTCTime)

Time in the past from which to export table data. The table export will be a snapshot of the table's state at this point in time.

exportTableToPointInTime_s3SseAlgorithm :: Lens' ExportTableToPointInTime (Maybe S3SseAlgorithm)

Type of encryption used on the bucket where export data will be stored. Valid values for S3SseAlgorithm are:

AES256 - server-side encryption with Amazon S3 managed keys

KMS - server-side encryption with AWS KMS managed keys

exportTableToPointInTime_s3Prefix :: Lens' ExportTableToPointInTime (Maybe Text)

The Amazon S3 bucket prefix to use as the file name and path of the exported snapshot.

exportTableToPointInTime_tableArn :: Lens' ExportTableToPointInTime Text

The Amazon Resource Name (ARN) associated with the table to export.

exportTableToPointInTime_s3Bucket :: Lens' ExportTableToPointInTime Text

The name of the Amazon S3 bucket to export the snapshot to.

Destructuring the Response

data ExportTableToPointInTimeResponse

See: newExportTableToPointInTimeResponse smart constructor.

Constructors

ExportTableToPointInTimeResponse'

Fields

exportDescription :: Maybe ExportDescription

Contains a description of the table export.

httpStatus :: Int

The response's http status code.

newExportTableToPointInTimeResponse

Arguments

:: Int
$sel:httpStatus:ExportTableToPointInTimeResponse'

-> ExportTableToPointInTimeResponse

Create a value of ExportTableToPointInTimeResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:exportDescription:ExportTableToPointInTimeResponse', exportTableToPointInTimeResponse_exportDescription - Contains a description of the table export.

$sel:httpStatus:ExportTableToPointInTimeResponse', exportTableToPointInTimeResponse_httpStatus - The response's http status code.

Response Lenses

exportTableToPointInTimeResponse_exportDescription :: Lens' ExportTableToPointInTimeResponse (Maybe ExportDescription)

Contains a description of the table export.

exportTableToPointInTimeResponse_httpStatus :: Lens' ExportTableToPointInTimeResponse Int

The response's http status code.