Amazonka.DynamoDB.RestoreTableFromBackup

Description

Creates a new table from an existing backup. Any number of users can execute up to 4 concurrent restores (any type of restore) in a given account.

You can call RestoreTableFromBackup at a maximum rate of 10 times per second.

You must manually set up the following on the restored table:

Auto scaling policies

IAM policies

Amazon CloudWatch metrics and alarms

Tags

Stream settings

Time to Live (TTL) settings

Creating a Request

data RestoreTableFromBackup

See: newRestoreTableFromBackup smart constructor.

Constructors

RestoreTableFromBackup'

Fields

billingModeOverride :: Maybe BillingMode

The billing mode of the restored table.

globalSecondaryIndexOverride :: Maybe [GlobalSecondaryIndex]

List of global secondary indexes for the restored table. The indexes provided should match existing secondary indexes. You can choose to exclude some or all of the indexes at the time of restore.

provisionedThroughputOverride :: Maybe ProvisionedThroughput

Provisioned throughput settings for the restored table.

sSESpecificationOverride :: Maybe SSESpecification

The new server-side encryption settings for the restored table.

localSecondaryIndexOverride :: Maybe [LocalSecondaryIndex]

List of local secondary indexes for the restored table. The indexes provided should match existing secondary indexes. You can choose to exclude some or all of the indexes at the time of restore.

targetTableName :: Text

The name of the new table to which the backup must be restored.

backupArn :: Text

The Amazon Resource Name (ARN) associated with the backup.

newRestoreTableFromBackup

Arguments

:: Text
$sel:targetTableName:RestoreTableFromBackup'

-> Text
$sel:backupArn:RestoreTableFromBackup'

-> RestoreTableFromBackup

Create a value of RestoreTableFromBackup with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:billingModeOverride:RestoreTableFromBackup', restoreTableFromBackup_billingModeOverride - The billing mode of the restored table.

$sel:globalSecondaryIndexOverride:RestoreTableFromBackup', restoreTableFromBackup_globalSecondaryIndexOverride - List of global secondary indexes for the restored table. The indexes provided should match existing secondary indexes. You can choose to exclude some or all of the indexes at the time of restore.

$sel:provisionedThroughputOverride:RestoreTableFromBackup', restoreTableFromBackup_provisionedThroughputOverride - Provisioned throughput settings for the restored table.

$sel:sSESpecificationOverride:RestoreTableFromBackup', restoreTableFromBackup_sSESpecificationOverride - The new server-side encryption settings for the restored table.

$sel:localSecondaryIndexOverride:RestoreTableFromBackup', restoreTableFromBackup_localSecondaryIndexOverride - List of local secondary indexes for the restored table. The indexes provided should match existing secondary indexes. You can choose to exclude some or all of the indexes at the time of restore.

$sel:targetTableName:RestoreTableFromBackup', restoreTableFromBackup_targetTableName - The name of the new table to which the backup must be restored.

$sel:backupArn:RestoreTableFromBackup', restoreTableFromBackup_backupArn - The Amazon Resource Name (ARN) associated with the backup.

Request Lenses

restoreTableFromBackup_billingModeOverride :: Lens' RestoreTableFromBackup (Maybe BillingMode)

The billing mode of the restored table.

restoreTableFromBackup_globalSecondaryIndexOverride :: Lens' RestoreTableFromBackup (Maybe [GlobalSecondaryIndex])

List of global secondary indexes for the restored table. The indexes provided should match existing secondary indexes. You can choose to exclude some or all of the indexes at the time of restore.

restoreTableFromBackup_provisionedThroughputOverride :: Lens' RestoreTableFromBackup (Maybe ProvisionedThroughput)

Provisioned throughput settings for the restored table.

restoreTableFromBackup_sSESpecificationOverride :: Lens' RestoreTableFromBackup (Maybe SSESpecification)

The new server-side encryption settings for the restored table.

restoreTableFromBackup_localSecondaryIndexOverride :: Lens' RestoreTableFromBackup (Maybe [LocalSecondaryIndex])

List of local secondary indexes for the restored table. The indexes provided should match existing secondary indexes. You can choose to exclude some or all of the indexes at the time of restore.

restoreTableFromBackup_targetTableName :: Lens' RestoreTableFromBackup Text

The name of the new table to which the backup must be restored.

restoreTableFromBackup_backupArn :: Lens' RestoreTableFromBackup Text

The Amazon Resource Name (ARN) associated with the backup.

Destructuring the Response

data RestoreTableFromBackupResponse

See: newRestoreTableFromBackupResponse smart constructor.

Constructors

RestoreTableFromBackupResponse'

Fields

tableDescription :: Maybe TableDescription

The description of the table created from an existing backup.

httpStatus :: Int

The response's http status code.

newRestoreTableFromBackupResponse

Arguments

:: Int
$sel:httpStatus:RestoreTableFromBackupResponse'

-> RestoreTableFromBackupResponse

Create a value of RestoreTableFromBackupResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:tableDescription:RestoreTableFromBackupResponse', restoreTableFromBackupResponse_tableDescription - The description of the table created from an existing backup.

$sel:httpStatus:RestoreTableFromBackupResponse', restoreTableFromBackupResponse_httpStatus - The response's http status code.

Response Lenses

restoreTableFromBackupResponse_tableDescription :: Lens' RestoreTableFromBackupResponse (Maybe TableDescription)

The description of the table created from an existing backup.

restoreTableFromBackupResponse_httpStatus :: Lens' RestoreTableFromBackupResponse Int

The response's http status code.