Amazonka.DynamoDB.Types.ContributorInsightsSummary

Description

Documentation

data ContributorInsightsSummary

Represents a Contributor Insights summary entry.

See: newContributorInsightsSummary smart constructor.

Constructors

ContributorInsightsSummary'

Fields

contributorInsightsStatus :: Maybe ContributorInsightsStatus

Describes the current status for contributor insights for the given table and index, if applicable.

tableName :: Maybe Text

Name of the table associated with the summary.

indexName :: Maybe Text

Name of the index associated with the summary, if any.

newContributorInsightsSummary :: ContributorInsightsSummary

Create a value of ContributorInsightsSummary with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:contributorInsightsStatus:ContributorInsightsSummary', contributorInsightsSummary_contributorInsightsStatus - Describes the current status for contributor insights for the given table and index, if applicable.

$sel:tableName:ContributorInsightsSummary', contributorInsightsSummary_tableName - Name of the table associated with the summary.

$sel:indexName:ContributorInsightsSummary', contributorInsightsSummary_indexName - Name of the index associated with the summary, if any.

contributorInsightsSummary_contributorInsightsStatus :: Lens' ContributorInsightsSummary (Maybe ContributorInsightsStatus)

Describes the current status for contributor insights for the given table and index, if applicable.

contributorInsightsSummary_tableName :: Lens' ContributorInsightsSummary (Maybe Text)

Name of the table associated with the summary.

contributorInsightsSummary_indexName :: Lens' ContributorInsightsSummary (Maybe Text)

Name of the index associated with the summary, if any.