Amazonka.DynamoDB.Types.ExportDescription

Description

Documentation

data ExportDescription

Represents the properties of the exported table.

See: newExportDescription smart constructor.

Constructors

ExportDescription'

Fields

s3BucketOwner :: Maybe Text

The ID of the AWS account that owns the bucket containing the export.

exportFormat :: Maybe ExportFormat

The format of the exported data. Valid values for ExportFormat are DYNAMODB_JSON or ION.

s3SseKmsKeyId :: Maybe Text

The ID of the AWS KMS managed key used to encrypt the S3 bucket where export data is stored (if applicable).

clientToken :: Maybe Text

The client token that was provided for the export task. A client token makes calls to ExportTableToPointInTimeInput idempotent, meaning that multiple identical calls have the same effect as one single call.

startTime :: Maybe POSIX

The time at which the export task began.

failureCode :: Maybe Text

Status code for the result of the failed export.

exportStatus :: Maybe ExportStatus

Export can be in one of the following states: IN_PROGRESS, COMPLETED, or FAILED.

failureMessage :: Maybe Text

Export failure reason description.

tableArn :: Maybe Text

The Amazon Resource Name (ARN) of the table that was exported.

billedSizeBytes :: Maybe Natural

The billable size of the table export.

exportArn :: Maybe Text

The Amazon Resource Name (ARN) of the table export.

exportTime :: Maybe POSIX

Point in time from which table data was exported.

s3SseAlgorithm :: Maybe S3SseAlgorithm

Type of encryption used on the bucket where export data is stored. Valid values for S3SseAlgorithm are:

AES256 - server-side encryption with Amazon S3 managed keys

KMS - server-side encryption with AWS KMS managed keys

endTime :: Maybe POSIX

The time at which the export task completed.

s3Prefix :: Maybe Text

The Amazon S3 bucket prefix used as the file name and path of the exported snapshot.

exportManifest :: Maybe Text

The name of the manifest file for the export task.

tableId :: Maybe Text

Unique ID of the table that was exported.

itemCount :: Maybe Natural

The number of items exported.

s3Bucket :: Maybe Text

The name of the Amazon S3 bucket containing the export.

newExportDescription :: ExportDescription

Create a value of ExportDescription with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:s3BucketOwner:ExportDescription', exportDescription_s3BucketOwner - The ID of the AWS account that owns the bucket containing the export.

$sel:exportFormat:ExportDescription', exportDescription_exportFormat - The format of the exported data. Valid values for ExportFormat are DYNAMODB_JSON or ION.

$sel:s3SseKmsKeyId:ExportDescription', exportDescription_s3SseKmsKeyId - The ID of the AWS KMS managed key used to encrypt the S3 bucket where export data is stored (if applicable).

$sel:clientToken:ExportDescription', exportDescription_clientToken - The client token that was provided for the export task. A client token makes calls to ExportTableToPointInTimeInput idempotent, meaning that multiple identical calls have the same effect as one single call.

$sel:startTime:ExportDescription', exportDescription_startTime - The time at which the export task began.

$sel:failureCode:ExportDescription', exportDescription_failureCode - Status code for the result of the failed export.

$sel:exportStatus:ExportDescription', exportDescription_exportStatus - Export can be in one of the following states: IN_PROGRESS, COMPLETED, or FAILED.

$sel:failureMessage:ExportDescription', exportDescription_failureMessage - Export failure reason description.

$sel:tableArn:ExportDescription', exportDescription_tableArn - The Amazon Resource Name (ARN) of the table that was exported.

$sel:billedSizeBytes:ExportDescription', exportDescription_billedSizeBytes - The billable size of the table export.

$sel:exportArn:ExportDescription', exportDescription_exportArn - The Amazon Resource Name (ARN) of the table export.

$sel:exportTime:ExportDescription', exportDescription_exportTime - Point in time from which table data was exported.

$sel:s3SseAlgorithm:ExportDescription', exportDescription_s3SseAlgorithm - Type of encryption used on the bucket where export data is stored. Valid values for S3SseAlgorithm are:

AES256 - server-side encryption with Amazon S3 managed keys

KMS - server-side encryption with AWS KMS managed keys

$sel:endTime:ExportDescription', exportDescription_endTime - The time at which the export task completed.

$sel:s3Prefix:ExportDescription', exportDescription_s3Prefix - The Amazon S3 bucket prefix used as the file name and path of the exported snapshot.

$sel:exportManifest:ExportDescription', exportDescription_exportManifest - The name of the manifest file for the export task.

$sel:tableId:ExportDescription', exportDescription_tableId - Unique ID of the table that was exported.

$sel:itemCount:ExportDescription', exportDescription_itemCount - The number of items exported.

$sel:s3Bucket:ExportDescription', exportDescription_s3Bucket - The name of the Amazon S3 bucket containing the export.

exportDescription_s3BucketOwner :: Lens' ExportDescription (Maybe Text)

The ID of the AWS account that owns the bucket containing the export.

exportDescription_exportFormat :: Lens' ExportDescription (Maybe ExportFormat)

The format of the exported data. Valid values for ExportFormat are DYNAMODB_JSON or ION.

exportDescription_s3SseKmsKeyId :: Lens' ExportDescription (Maybe Text)

The ID of the AWS KMS managed key used to encrypt the S3 bucket where export data is stored (if applicable).

exportDescription_clientToken :: Lens' ExportDescription (Maybe Text)

The client token that was provided for the export task. A client token makes calls to ExportTableToPointInTimeInput idempotent, meaning that multiple identical calls have the same effect as one single call.

exportDescription_startTime :: Lens' ExportDescription (Maybe UTCTime)

The time at which the export task began.

exportDescription_failureCode :: Lens' ExportDescription (Maybe Text)

Status code for the result of the failed export.

exportDescription_exportStatus :: Lens' ExportDescription (Maybe ExportStatus)

Export can be in one of the following states: IN_PROGRESS, COMPLETED, or FAILED.

exportDescription_failureMessage :: Lens' ExportDescription (Maybe Text)

Export failure reason description.

exportDescription_tableArn :: Lens' ExportDescription (Maybe Text)

The Amazon Resource Name (ARN) of the table that was exported.

exportDescription_billedSizeBytes :: Lens' ExportDescription (Maybe Natural)

The billable size of the table export.

exportDescription_exportArn :: Lens' ExportDescription (Maybe Text)

The Amazon Resource Name (ARN) of the table export.

exportDescription_exportTime :: Lens' ExportDescription (Maybe UTCTime)

Point in time from which table data was exported.

exportDescription_s3SseAlgorithm :: Lens' ExportDescription (Maybe S3SseAlgorithm)

Type of encryption used on the bucket where export data is stored. Valid values for S3SseAlgorithm are:

AES256 - server-side encryption with Amazon S3 managed keys

KMS - server-side encryption with AWS KMS managed keys

exportDescription_endTime :: Lens' ExportDescription (Maybe UTCTime)

The time at which the export task completed.

exportDescription_s3Prefix :: Lens' ExportDescription (Maybe Text)

The Amazon S3 bucket prefix used as the file name and path of the exported snapshot.

exportDescription_exportManifest :: Lens' ExportDescription (Maybe Text)

The name of the manifest file for the export task.

exportDescription_tableId :: Lens' ExportDescription (Maybe Text)

Unique ID of the table that was exported.

exportDescription_itemCount :: Lens' ExportDescription (Maybe Natural)

The number of items exported.

exportDescription_s3Bucket :: Lens' ExportDescription (Maybe Text)

The name of the Amazon S3 bucket containing the export.