Amazonka.DynamoDB.Types.GlobalSecondaryIndexAutoScalingUpdate

Description

Documentation

data GlobalSecondaryIndexAutoScalingUpdate

Represents the auto scaling settings of a global secondary index for a global table that will be modified.

See: newGlobalSecondaryIndexAutoScalingUpdate smart constructor.

Constructors

GlobalSecondaryIndexAutoScalingUpdate'

Fields

provisionedWriteCapacityAutoScalingUpdate :: Maybe AutoScalingSettingsUpdate

indexName :: Maybe Text

The name of the global secondary index.

newGlobalSecondaryIndexAutoScalingUpdate :: GlobalSecondaryIndexAutoScalingUpdate

Create a value of GlobalSecondaryIndexAutoScalingUpdate with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:provisionedWriteCapacityAutoScalingUpdate:GlobalSecondaryIndexAutoScalingUpdate', globalSecondaryIndexAutoScalingUpdate_provisionedWriteCapacityAutoScalingUpdate - Undocumented member.

$sel:indexName:GlobalSecondaryIndexAutoScalingUpdate', globalSecondaryIndexAutoScalingUpdate_indexName - The name of the global secondary index.

globalSecondaryIndexAutoScalingUpdate_provisionedWriteCapacityAutoScalingUpdate :: Lens' GlobalSecondaryIndexAutoScalingUpdate (Maybe AutoScalingSettingsUpdate)

Undocumented member.

globalSecondaryIndexAutoScalingUpdate_indexName :: Lens' GlobalSecondaryIndexAutoScalingUpdate (Maybe Text)

The name of the global secondary index.