Amazonka.DynamoDB.Types.GlobalSecondaryIndexUpdate

Description

Documentation

data GlobalSecondaryIndexUpdate

Represents one of the following:

A new global secondary index to be added to an existing table.

New provisioned throughput parameters for an existing global secondary index.

An existing global secondary index to be removed from an existing table.

See: newGlobalSecondaryIndexUpdate smart constructor.

Constructors

GlobalSecondaryIndexUpdate'

Fields

create :: Maybe CreateGlobalSecondaryIndexAction

The parameters required for creating a global secondary index on an existing table:

IndexName

KeySchema

AttributeDefinitions

Projection

ProvisionedThroughput

delete' :: Maybe DeleteGlobalSecondaryIndexAction

The name of an existing global secondary index to be removed.

update :: Maybe UpdateGlobalSecondaryIndexAction

The name of an existing global secondary index, along with new provisioned throughput settings to be applied to that index.

newGlobalSecondaryIndexUpdate :: GlobalSecondaryIndexUpdate

Create a value of GlobalSecondaryIndexUpdate with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:create:GlobalSecondaryIndexUpdate', globalSecondaryIndexUpdate_create - The parameters required for creating a global secondary index on an existing table:

IndexName

KeySchema

AttributeDefinitions

Projection

ProvisionedThroughput

$sel:delete':GlobalSecondaryIndexUpdate', globalSecondaryIndexUpdate_delete - The name of an existing global secondary index to be removed.

$sel:update:GlobalSecondaryIndexUpdate', globalSecondaryIndexUpdate_update - The name of an existing global secondary index, along with new provisioned throughput settings to be applied to that index.

globalSecondaryIndexUpdate_create :: Lens' GlobalSecondaryIndexUpdate (Maybe CreateGlobalSecondaryIndexAction)

The parameters required for creating a global secondary index on an existing table:

IndexName

KeySchema

AttributeDefinitions

Projection

ProvisionedThroughput

globalSecondaryIndexUpdate_delete :: Lens' GlobalSecondaryIndexUpdate (Maybe DeleteGlobalSecondaryIndexAction)

The name of an existing global secondary index to be removed.

globalSecondaryIndexUpdate_update :: Lens' GlobalSecondaryIndexUpdate (Maybe UpdateGlobalSecondaryIndexAction)

The name of an existing global secondary index, along with new provisioned throughput settings to be applied to that index.