Amazonka.DynamoDB.Types.TransactWriteItem

Description

Documentation

data TransactWriteItem

A list of requests that can perform update, put, delete, or check operations on multiple items in one or more tables atomically.

See: newTransactWriteItem smart constructor.

Constructors

TransactWriteItem'

Fields

conditionCheck :: Maybe ConditionCheck

A request to perform a check item operation.

put :: Maybe Put

A request to perform a PutItem operation.

delete' :: Maybe Delete

A request to perform a DeleteItem operation.

update :: Maybe Update

A request to perform an UpdateItem operation.

newTransactWriteItem :: TransactWriteItem

Create a value of TransactWriteItem with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:conditionCheck:TransactWriteItem', transactWriteItem_conditionCheck - A request to perform a check item operation.

$sel:put:TransactWriteItem', transactWriteItem_put - A request to perform a PutItem operation.

$sel:delete':TransactWriteItem', transactWriteItem_delete - A request to perform a DeleteItem operation.

$sel:update:TransactWriteItem', transactWriteItem_update - A request to perform an UpdateItem operation.

transactWriteItem_conditionCheck :: Lens' TransactWriteItem (Maybe ConditionCheck)

A request to perform a check item operation.

transactWriteItem_put :: Lens' TransactWriteItem (Maybe Put)

A request to perform a PutItem operation.

transactWriteItem_delete :: Lens' TransactWriteItem (Maybe Delete)

A request to perform a DeleteItem operation.

transactWriteItem_update :: Lens' TransactWriteItem (Maybe Update)

A request to perform an UpdateItem operation.