Amazonka.DynamoDB.UpdateGlobalTable

Description

Adds or removes replicas in the specified global table. The global table must already exist to be able to use this operation. Any replica to be added must be empty, have the same name as the global table, have the same key schema, have DynamoDB Streams enabled, and have the same provisioned and maximum write capacity units.

Although you can use UpdateGlobalTable to add replicas and remove replicas in a single request, for simplicity we recommend that you issue separate requests for adding or removing replicas.

If global secondary indexes are specified, then the following conditions must also be met:

The global secondary indexes must have the same name.

The global secondary indexes must have the same hash key and sort key (if present).

The global secondary indexes must have the same provisioned and maximum write capacity units.

Creating a Request

data UpdateGlobalTable

See: newUpdateGlobalTable smart constructor.

Constructors

UpdateGlobalTable'

Fields

globalTableName :: Text

The global table name.

replicaUpdates :: [ReplicaUpdate]

A list of Regions that should be added or removed from the global table.

newUpdateGlobalTable

Arguments

:: Text
$sel:globalTableName:UpdateGlobalTable'

-> UpdateGlobalTable

Create a value of UpdateGlobalTable with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:globalTableName:UpdateGlobalTable', updateGlobalTable_globalTableName - The global table name.

$sel:replicaUpdates:UpdateGlobalTable', updateGlobalTable_replicaUpdates - A list of Regions that should be added or removed from the global table.

Request Lenses

updateGlobalTable_globalTableName :: Lens' UpdateGlobalTable Text

The global table name.

updateGlobalTable_replicaUpdates :: Lens' UpdateGlobalTable [ReplicaUpdate]

A list of Regions that should be added or removed from the global table.

Destructuring the Response

data UpdateGlobalTableResponse

See: newUpdateGlobalTableResponse smart constructor.

Constructors

UpdateGlobalTableResponse'

Fields

globalTableDescription :: Maybe GlobalTableDescription

Contains the details of the global table.

httpStatus :: Int

The response's http status code.

newUpdateGlobalTableResponse

Arguments

:: Int
$sel:httpStatus:UpdateGlobalTableResponse'

-> UpdateGlobalTableResponse

Create a value of UpdateGlobalTableResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:globalTableDescription:UpdateGlobalTableResponse', updateGlobalTableResponse_globalTableDescription - Contains the details of the global table.

$sel:httpStatus:UpdateGlobalTableResponse', updateGlobalTableResponse_httpStatus - The response's http status code.

Response Lenses

updateGlobalTableResponse_globalTableDescription :: Lens' UpdateGlobalTableResponse (Maybe GlobalTableDescription)

Contains the details of the global table.

updateGlobalTableResponse_httpStatus :: Lens' UpdateGlobalTableResponse Int

The response's http status code.