Amazonka.DynamoDB.UpdateGlobalTableSettings

Description

Updates settings for a global table.

Creating a Request

data UpdateGlobalTableSettings

See: newUpdateGlobalTableSettings smart constructor.

Constructors

UpdateGlobalTableSettings'

Fields

globalTableProvisionedWriteCapacityAutoScalingSettingsUpdate :: Maybe AutoScalingSettingsUpdate

Auto scaling settings for managing provisioned write capacity for the global table.

globalTableBillingMode :: Maybe BillingMode

The billing mode of the global table. If GlobalTableBillingMode is not specified, the global table defaults to PROVISIONED capacity billing mode.

PROVISIONED - We recommend using PROVISIONED for predictable workloads. PROVISIONED sets the billing mode to Provisioned Mode.

PAY_PER_REQUEST - We recommend using PAY_PER_REQUEST for unpredictable workloads. PAY_PER_REQUEST sets the billing mode to On-Demand Mode.

globalTableProvisionedWriteCapacityUnits :: Maybe Natural

The maximum number of writes consumed per second before DynamoDB returns a ThrottlingException.

replicaSettingsUpdate :: Maybe (NonEmpty ReplicaSettingsUpdate)

Represents the settings for a global table in a Region that will be modified.

globalTableGlobalSecondaryIndexSettingsUpdate :: Maybe (NonEmpty GlobalTableGlobalSecondaryIndexSettingsUpdate)

Represents the settings of a global secondary index for a global table that will be modified.

globalTableName :: Text

The name of the global table

newUpdateGlobalTableSettings

Arguments

:: Text
$sel:globalTableName:UpdateGlobalTableSettings'

-> UpdateGlobalTableSettings

Create a value of UpdateGlobalTableSettings with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:globalTableProvisionedWriteCapacityAutoScalingSettingsUpdate:UpdateGlobalTableSettings', updateGlobalTableSettings_globalTableProvisionedWriteCapacityAutoScalingSettingsUpdate - Auto scaling settings for managing provisioned write capacity for the global table.

$sel:globalTableBillingMode:UpdateGlobalTableSettings', updateGlobalTableSettings_globalTableBillingMode - The billing mode of the global table. If GlobalTableBillingMode is not specified, the global table defaults to PROVISIONED capacity billing mode.

PROVISIONED - We recommend using PROVISIONED for predictable workloads. PROVISIONED sets the billing mode to Provisioned Mode.

PAY_PER_REQUEST - We recommend using PAY_PER_REQUEST for unpredictable workloads. PAY_PER_REQUEST sets the billing mode to On-Demand Mode.

$sel:globalTableProvisionedWriteCapacityUnits:UpdateGlobalTableSettings', updateGlobalTableSettings_globalTableProvisionedWriteCapacityUnits - The maximum number of writes consumed per second before DynamoDB returns a ThrottlingException.

$sel:replicaSettingsUpdate:UpdateGlobalTableSettings', updateGlobalTableSettings_replicaSettingsUpdate - Represents the settings for a global table in a Region that will be modified.

$sel:globalTableGlobalSecondaryIndexSettingsUpdate:UpdateGlobalTableSettings', updateGlobalTableSettings_globalTableGlobalSecondaryIndexSettingsUpdate - Represents the settings of a global secondary index for a global table that will be modified.

$sel:globalTableName:UpdateGlobalTableSettings', updateGlobalTableSettings_globalTableName - The name of the global table

Request Lenses

updateGlobalTableSettings_globalTableProvisionedWriteCapacityAutoScalingSettingsUpdate :: Lens' UpdateGlobalTableSettings (Maybe AutoScalingSettingsUpdate)

Auto scaling settings for managing provisioned write capacity for the global table.

updateGlobalTableSettings_globalTableBillingMode :: Lens' UpdateGlobalTableSettings (Maybe BillingMode)

The billing mode of the global table. If GlobalTableBillingMode is not specified, the global table defaults to PROVISIONED capacity billing mode.

PROVISIONED - We recommend using PROVISIONED for predictable workloads. PROVISIONED sets the billing mode to Provisioned Mode.

PAY_PER_REQUEST - We recommend using PAY_PER_REQUEST for unpredictable workloads. PAY_PER_REQUEST sets the billing mode to On-Demand Mode.

updateGlobalTableSettings_globalTableProvisionedWriteCapacityUnits :: Lens' UpdateGlobalTableSettings (Maybe Natural)

The maximum number of writes consumed per second before DynamoDB returns a ThrottlingException.

updateGlobalTableSettings_replicaSettingsUpdate :: Lens' UpdateGlobalTableSettings (Maybe (NonEmpty ReplicaSettingsUpdate))

Represents the settings for a global table in a Region that will be modified.

updateGlobalTableSettings_globalTableGlobalSecondaryIndexSettingsUpdate :: Lens' UpdateGlobalTableSettings (Maybe (NonEmpty GlobalTableGlobalSecondaryIndexSettingsUpdate))

Represents the settings of a global secondary index for a global table that will be modified.

updateGlobalTableSettings_globalTableName :: Lens' UpdateGlobalTableSettings Text

The name of the global table

Destructuring the Response

data UpdateGlobalTableSettingsResponse

See: newUpdateGlobalTableSettingsResponse smart constructor.

Constructors

UpdateGlobalTableSettingsResponse'

Fields

replicaSettings :: Maybe [ReplicaSettingsDescription]

The Region-specific settings for the global table.

globalTableName :: Maybe Text

The name of the global table.

httpStatus :: Int

The response's http status code.

newUpdateGlobalTableSettingsResponse

Arguments

:: Int
$sel:httpStatus:UpdateGlobalTableSettingsResponse'

-> UpdateGlobalTableSettingsResponse

Create a value of UpdateGlobalTableSettingsResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:replicaSettings:UpdateGlobalTableSettingsResponse', updateGlobalTableSettingsResponse_replicaSettings - The Region-specific settings for the global table.

$sel:globalTableName:UpdateGlobalTableSettings', updateGlobalTableSettingsResponse_globalTableName - The name of the global table.

$sel:httpStatus:UpdateGlobalTableSettingsResponse', updateGlobalTableSettingsResponse_httpStatus - The response's http status code.

Response Lenses

updateGlobalTableSettingsResponse_replicaSettings :: Lens' UpdateGlobalTableSettingsResponse (Maybe [ReplicaSettingsDescription])

The Region-specific settings for the global table.

updateGlobalTableSettingsResponse_globalTableName :: Lens' UpdateGlobalTableSettingsResponse (Maybe Text)

The name of the global table.

updateGlobalTableSettingsResponse_httpStatus :: Lens' UpdateGlobalTableSettingsResponse Int

The response's http status code.