Amazonka.DynamoDB.UpdateTable

Description

Modifies the provisioned throughput settings, global secondary indexes, or DynamoDB Streams settings for a given table.

You can only perform one of the following operations at once:

Modify the provisioned throughput settings of the table.

Enable or disable DynamoDB Streams on the table.

Remove a global secondary index from the table.

Create a new global secondary index on the table. After the index begins backfilling, you can use UpdateTable to perform other operations.

UpdateTable is an asynchronous operation; while it is executing, the table status changes from ACTIVE to UPDATING. While it is UPDATING, you cannot issue another UpdateTable request. When the table returns to the ACTIVE state, the UpdateTable operation is complete.

Creating a Request

data UpdateTable

Represents the input of an UpdateTable operation.

See: newUpdateTable smart constructor.

Constructors

UpdateTable'

Fields

attributeDefinitions :: Maybe [AttributeDefinition]

An array of attributes that describe the key schema for the table and indexes. If you are adding a new global secondary index to the table, AttributeDefinitions must include the key element(s) of the new index.

provisionedThroughput :: Maybe ProvisionedThroughput

The new provisioned throughput settings for the specified table or index.

sSESpecification :: Maybe SSESpecification

The new server-side encryption settings for the specified table.

replicaUpdates :: Maybe (NonEmpty ReplicationGroupUpdate)

A list of replica update actions (create, delete, or update) for the table.

This property only applies to Version 2019.11.21 of global tables.

globalSecondaryIndexUpdates :: Maybe [GlobalSecondaryIndexUpdate]

An array of one or more global secondary indexes for the table. For each index in the array, you can request one action:

Create - add a new global secondary index to the table.

Update - modify the provisioned throughput settings of an existing global secondary index.

Delete - remove a global secondary index from the table.

You can create or delete only one global secondary index per UpdateTable operation.

For more information, see Managing Global Secondary Indexes in the Amazon DynamoDB Developer Guide.

billingMode :: Maybe BillingMode

Controls how you are charged for read and write throughput and how you manage capacity. When switching from pay-per-request to provisioned capacity, initial provisioned capacity values must be set. The initial provisioned capacity values are estimated based on the consumed read and write capacity of your table and global secondary indexes over the past 30 minutes.

PROVISIONED - We recommend using PROVISIONED for predictable workloads. PROVISIONED sets the billing mode to Provisioned Mode.

PAY_PER_REQUEST - We recommend using PAY_PER_REQUEST for unpredictable workloads. PAY_PER_REQUEST sets the billing mode to On-Demand Mode.

streamSpecification :: Maybe StreamSpecification

Represents the DynamoDB Streams configuration for the table.

You receive a ResourceInUseException if you try to enable a stream on a table that already has a stream, or if you try to disable a stream on a table that doesn't have a stream.

tableName :: Text

The name of the table to be updated.

newUpdateTable

Arguments

:: Text
$sel:tableName:UpdateTable'

-> UpdateTable

Create a value of UpdateTable with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:attributeDefinitions:UpdateTable', updateTable_attributeDefinitions - An array of attributes that describe the key schema for the table and indexes. If you are adding a new global secondary index to the table, AttributeDefinitions must include the key element(s) of the new index.

$sel:provisionedThroughput:UpdateTable', updateTable_provisionedThroughput - The new provisioned throughput settings for the specified table or index.

$sel:sSESpecification:UpdateTable', updateTable_sSESpecification - The new server-side encryption settings for the specified table.

$sel:replicaUpdates:UpdateTable', updateTable_replicaUpdates - A list of replica update actions (create, delete, or update) for the table.

This property only applies to Version 2019.11.21 of global tables.

$sel:globalSecondaryIndexUpdates:UpdateTable', updateTable_globalSecondaryIndexUpdates - An array of one or more global secondary indexes for the table. For each index in the array, you can request one action:

Create - add a new global secondary index to the table.

Update - modify the provisioned throughput settings of an existing global secondary index.

Delete - remove a global secondary index from the table.

You can create or delete only one global secondary index per UpdateTable operation.

For more information, see Managing Global Secondary Indexes in the Amazon DynamoDB Developer Guide.

$sel:billingMode:UpdateTable', updateTable_billingMode - Controls how you are charged for read and write throughput and how you manage capacity. When switching from pay-per-request to provisioned capacity, initial provisioned capacity values must be set. The initial provisioned capacity values are estimated based on the consumed read and write capacity of your table and global secondary indexes over the past 30 minutes.

PROVISIONED - We recommend using PROVISIONED for predictable workloads. PROVISIONED sets the billing mode to Provisioned Mode.

PAY_PER_REQUEST - We recommend using PAY_PER_REQUEST for unpredictable workloads. PAY_PER_REQUEST sets the billing mode to On-Demand Mode.

$sel:streamSpecification:UpdateTable', updateTable_streamSpecification - Represents the DynamoDB Streams configuration for the table.

You receive a ResourceInUseException if you try to enable a stream on a table that already has a stream, or if you try to disable a stream on a table that doesn't have a stream.

$sel:tableName:UpdateTable', updateTable_tableName - The name of the table to be updated.

Request Lenses

updateTable_attributeDefinitions :: Lens' UpdateTable (Maybe [AttributeDefinition])

An array of attributes that describe the key schema for the table and indexes. If you are adding a new global secondary index to the table, AttributeDefinitions must include the key element(s) of the new index.

updateTable_provisionedThroughput :: Lens' UpdateTable (Maybe ProvisionedThroughput)

The new provisioned throughput settings for the specified table or index.

updateTable_sSESpecification :: Lens' UpdateTable (Maybe SSESpecification)

The new server-side encryption settings for the specified table.

updateTable_replicaUpdates :: Lens' UpdateTable (Maybe (NonEmpty ReplicationGroupUpdate))

A list of replica update actions (create, delete, or update) for the table.

This property only applies to Version 2019.11.21 of global tables.

updateTable_globalSecondaryIndexUpdates :: Lens' UpdateTable (Maybe [GlobalSecondaryIndexUpdate])

An array of one or more global secondary indexes for the table. For each index in the array, you can request one action:

Create - add a new global secondary index to the table.

Update - modify the provisioned throughput settings of an existing global secondary index.

Delete - remove a global secondary index from the table.

You can create or delete only one global secondary index per UpdateTable operation.

For more information, see Managing Global Secondary Indexes in the Amazon DynamoDB Developer Guide.

updateTable_billingMode :: Lens' UpdateTable (Maybe BillingMode)

Controls how you are charged for read and write throughput and how you manage capacity. When switching from pay-per-request to provisioned capacity, initial provisioned capacity values must be set. The initial provisioned capacity values are estimated based on the consumed read and write capacity of your table and global secondary indexes over the past 30 minutes.

PROVISIONED - We recommend using PROVISIONED for predictable workloads. PROVISIONED sets the billing mode to Provisioned Mode.

PAY_PER_REQUEST - We recommend using PAY_PER_REQUEST for unpredictable workloads. PAY_PER_REQUEST sets the billing mode to On-Demand Mode.

updateTable_streamSpecification :: Lens' UpdateTable (Maybe StreamSpecification)

Represents the DynamoDB Streams configuration for the table.

You receive a ResourceInUseException if you try to enable a stream on a table that already has a stream, or if you try to disable a stream on a table that doesn't have a stream.

updateTable_tableName :: Lens' UpdateTable Text

The name of the table to be updated.

Destructuring the Response

data UpdateTableResponse

Represents the output of an UpdateTable operation.

See: newUpdateTableResponse smart constructor.

Constructors

UpdateTableResponse'

Fields

tableDescription :: Maybe TableDescription

Represents the properties of the table.

httpStatus :: Int

The response's http status code.

newUpdateTableResponse

Arguments

:: Int
$sel:httpStatus:UpdateTableResponse'

-> UpdateTableResponse

Create a value of UpdateTableResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:tableDescription:UpdateTableResponse', updateTableResponse_tableDescription - Represents the properties of the table.

$sel:httpStatus:UpdateTableResponse', updateTableResponse_httpStatus - The response's http status code.

Response Lenses

updateTableResponse_tableDescription :: Lens' UpdateTableResponse (Maybe TableDescription)

Represents the properties of the table.

updateTableResponse_httpStatus :: Lens' UpdateTableResponse Int

The response's http status code.