Amazonka.DynamoDBStreams

Description

Derived from API version 2012-08-10 of the AWS service descriptions, licensed under Apache 2.0.

Amazon DynamoDB

Amazon DynamoDB Streams provides API actions for accessing streams and processing stream records. To learn more about application development with Streams, see Capturing Table Activity with DynamoDB Streams in the Amazon DynamoDB Developer Guide.

Service Configuration

defaultService :: Service

API version 2012-08-10 of the Amazon DynamoDB Streams SDK configuration.

Errors

Error matchers are designed for use with the functions provided by Control.Exception.Lens. This allows catching (and rethrowing) service specific errors returned by DynamoDBStreams.

ExpiredIteratorException

_ExpiredIteratorException :: AsError a => Getting (First ServiceError) a ServiceError

The shard iterator has expired and can no longer be used to retrieve stream records. A shard iterator expires 15 minutes after it is retrieved using the GetShardIterator action.

InternalServerError

_InternalServerError :: AsError a => Getting (First ServiceError) a ServiceError

An error occurred on the server side.

TrimmedDataAccessException

_TrimmedDataAccessException :: AsError a => Getting (First ServiceError) a ServiceError

The operation attempted to read past the oldest stream record in a shard.

In DynamoDB Streams, there is a 24 hour limit on data retention. Stream records whose age exceeds this limit are subject to removal (trimming) from the stream. You might receive a TrimmedDataAccessException if:

You request a shard iterator with a sequence number older than the trim point (24 hours).

You obtain a shard iterator, but before you use the iterator in a GetRecords request, a stream record in the shard exceeds the 24 hour period and is trimmed. This causes the iterator to access a record that no longer exists.

ResourceNotFoundException

_ResourceNotFoundException :: AsError a => Getting (First ServiceError) a ServiceError

The operation tried to access a nonexistent table or index. The resource might not be specified correctly, or its status might not be ACTIVE.

LimitExceededException

_LimitExceededException :: AsError a => Getting (First ServiceError) a ServiceError

There is no limit to the number of daily on-demand backups that can be taken.

Up to 50 simultaneous table operations are allowed per account. These operations include CreateTable, UpdateTable, DeleteTable,UpdateTimeToLive, RestoreTableFromBackup, and RestoreTableToPointInTime.

The only exception is when you are creating a table with one or more secondary indexes. You can have up to 25 such requests running at a time; however, if the table or index specifications are complex, DynamoDB might temporarily reduce the number of concurrent operations.

There is a soft account quota of 256 tables.

Waiters

Waiters poll by repeatedly sending a request until some remote success condition configured by the Wait specification is fulfilled. The Wait specification determines how many attempts should be made, in addition to delay and retry strategies.

Operations

Some AWS operations return results that are incomplete and require subsequent requests in order to obtain the entire result set. The process of sending subsequent requests to continue where a previous request left off is called pagination. For example, the ListObjects operation of Amazon S3 returns up to 1000 objects at a time, and you must send subsequent requests with the appropriate Marker in order to retrieve the next page of results.

Operations that have an AWSPager instance can transparently perform subsequent requests, correctly setting Markers and other request facets to iterate through the entire result set of a truncated API operation. Operations which support this have an additional note in the documentation.

Many operations have the ability to filter results on the server side. See the individual operation parameters for details.

GetShardIterator

data GetShardIterator

Represents the input of a GetShardIterator operation.

See: newGetShardIterator smart constructor.

Constructors

GetShardIterator' (Maybe Text) Text Text ShardIteratorType

newGetShardIterator

Arguments

:: Text
$sel:streamArn:GetShardIterator'

-> Text
$sel:shardId:GetShardIterator'

-> ShardIteratorType
$sel:shardIteratorType:GetShardIterator'

-> GetShardIterator

Create a value of GetShardIterator with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:sequenceNumber:GetShardIterator', getShardIterator_sequenceNumber - The sequence number of a stream record in the shard from which to start reading.

$sel:streamArn:GetShardIterator', getShardIterator_streamArn - The Amazon Resource Name (ARN) for the stream.

$sel:shardId:GetShardIterator', getShardIterator_shardId - The identifier of the shard. The iterator will be returned for this shard ID.

$sel:shardIteratorType:GetShardIterator', getShardIterator_shardIteratorType - Determines how the shard iterator is used to start reading stream records from the shard:

AT_SEQUENCE_NUMBER - Start reading exactly from the position denoted by a specific sequence number.

AFTER_SEQUENCE_NUMBER - Start reading right after the position denoted by a specific sequence number.

TRIM_HORIZON - Start reading at the last (untrimmed) stream record, which is the oldest record in the shard. In DynamoDB Streams, there is a 24 hour limit on data retention. Stream records whose age exceeds this limit are subject to removal (trimming) from the stream.

LATEST - Start reading just after the most recent stream record in the shard, so that you always read the most recent data in the shard.

data GetShardIteratorResponse

Represents the output of a GetShardIterator operation.

See: newGetShardIteratorResponse smart constructor.

Constructors

GetShardIteratorResponse' (Maybe Text) Int

newGetShardIteratorResponse

Arguments

:: Int
$sel:httpStatus:GetShardIteratorResponse'

-> GetShardIteratorResponse

Create a value of GetShardIteratorResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:shardIterator:GetShardIteratorResponse', getShardIteratorResponse_shardIterator - The position in the shard from which to start reading stream records sequentially. A shard iterator specifies this position using the sequence number of a stream record in a shard.

$sel:httpStatus:GetShardIteratorResponse', getShardIteratorResponse_httpStatus - The response's http status code.

GetRecords

data GetRecords

Represents the input of a GetRecords operation.

See: newGetRecords smart constructor.

Constructors

GetRecords' (Maybe Natural) Text

newGetRecords

Arguments

:: Text
$sel:shardIterator:GetRecords'

-> GetRecords

Create a value of GetRecords with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:limit:GetRecords', getRecords_limit - The maximum number of records to return from the shard. The upper limit is 1000.

$sel:shardIterator:GetRecords', getRecords_shardIterator - A shard iterator that was retrieved from a previous GetShardIterator operation. This iterator can be used to access the stream records in this shard.

data GetRecordsResponse

Represents the output of a GetRecords operation.

See: newGetRecordsResponse smart constructor.

Constructors

GetRecordsResponse' (Maybe [Record]) (Maybe Text) Int

newGetRecordsResponse

Arguments

:: Int
$sel:httpStatus:GetRecordsResponse'

-> GetRecordsResponse

Create a value of GetRecordsResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:records:GetRecordsResponse', getRecordsResponse_records - The stream records from the shard, which were retrieved using the shard iterator.

$sel:nextShardIterator:GetRecordsResponse', getRecordsResponse_nextShardIterator - The next position in the shard from which to start sequentially reading stream records. If set to null, the shard has been closed and the requested iterator will not return any more data.

$sel:httpStatus:GetRecordsResponse', getRecordsResponse_httpStatus - The response's http status code.

ListStreams

data ListStreams

Represents the input of a ListStreams operation.

See: newListStreams smart constructor.

Constructors

ListStreams' (Maybe Text) (Maybe Natural) (Maybe Text)

newListStreams :: ListStreams

Create a value of ListStreams with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:exclusiveStartStreamArn:ListStreams', listStreams_exclusiveStartStreamArn - The ARN (Amazon Resource Name) of the first item that this operation will evaluate. Use the value that was returned for LastEvaluatedStreamArn in the previous operation.

$sel:limit:ListStreams', listStreams_limit - The maximum number of streams to return. The upper limit is 100.

$sel:tableName:ListStreams', listStreams_tableName - If this parameter is provided, then only the streams associated with this table name are returned.

data ListStreamsResponse

Represents the output of a ListStreams operation.

See: newListStreamsResponse smart constructor.

Constructors

ListStreamsResponse' (Maybe Text) (Maybe [Stream]) Int

newListStreamsResponse

Arguments

:: Int
$sel:httpStatus:ListStreamsResponse'

-> ListStreamsResponse

Create a value of ListStreamsResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:lastEvaluatedStreamArn:ListStreamsResponse', listStreamsResponse_lastEvaluatedStreamArn - The stream ARN of the item where the operation stopped, inclusive of the previous result set. Use this value to start a new operation, excluding this value in the new request.

If LastEvaluatedStreamArn is empty, then the "last page" of results has been processed and there is no more data to be retrieved.

If LastEvaluatedStreamArn is not empty, it does not necessarily mean that there is more data in the result set. The only way to know when you have reached the end of the result set is when LastEvaluatedStreamArn is empty.

$sel:streams:ListStreamsResponse', listStreamsResponse_streams - A list of stream descriptors associated with the current account and endpoint.

$sel:httpStatus:ListStreamsResponse', listStreamsResponse_httpStatus - The response's http status code.

DescribeStream

data DescribeStream

Represents the input of a DescribeStream operation.

See: newDescribeStream smart constructor.

Constructors

DescribeStream' (Maybe Text) (Maybe Natural) Text

newDescribeStream

Arguments

:: Text
$sel:streamArn:DescribeStream'

-> DescribeStream

Create a value of DescribeStream with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:exclusiveStartShardId:DescribeStream', describeStream_exclusiveStartShardId - The shard ID of the first item that this operation will evaluate. Use the value that was returned for LastEvaluatedShardId in the previous operation.

$sel:limit:DescribeStream', describeStream_limit - The maximum number of shard objects to return. The upper limit is 100.

$sel:streamArn:DescribeStream', describeStream_streamArn - The Amazon Resource Name (ARN) for the stream.

data DescribeStreamResponse

Represents the output of a DescribeStream operation.

See: newDescribeStreamResponse smart constructor.

Constructors

DescribeStreamResponse' (Maybe StreamDescription) Int

newDescribeStreamResponse

Arguments

:: Int
$sel:httpStatus:DescribeStreamResponse'

-> DescribeStreamResponse

Create a value of DescribeStreamResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:streamDescription:DescribeStreamResponse', describeStreamResponse_streamDescription - A complete description of the stream, including its creation date and time, the DynamoDB table associated with the stream, the shard IDs within the stream, and the beginning and ending sequence numbers of stream records within the shards.

$sel:httpStatus:DescribeStreamResponse', describeStreamResponse_httpStatus - The response's http status code.

Types

KeyType

newtype KeyType

Constructors

KeyType'

Fields

fromKeyType :: Text

Bundled Patterns

pattern KeyType_HASH :: KeyType

pattern KeyType_RANGE :: KeyType

OperationType

newtype OperationType

Constructors

OperationType'

Fields

fromOperationType :: Text

Bundled Patterns

pattern OperationType_INSERT :: OperationType

pattern OperationType_MODIFY :: OperationType

pattern OperationType_REMOVE :: OperationType

ShardIteratorType

newtype ShardIteratorType

Constructors

ShardIteratorType'

Fields

fromShardIteratorType :: Text

Bundled Patterns

pattern ShardIteratorType_AFTER_SEQUENCE_NUMBER :: ShardIteratorType

pattern ShardIteratorType_AT_SEQUENCE_NUMBER :: ShardIteratorType

pattern ShardIteratorType_LATEST :: ShardIteratorType

pattern ShardIteratorType_TRIM_HORIZON :: ShardIteratorType

StreamStatus

newtype StreamStatus

Constructors

StreamStatus'

Fields

fromStreamStatus :: Text

Bundled Patterns

pattern StreamStatus_DISABLED :: StreamStatus

pattern StreamStatus_DISABLING :: StreamStatus

pattern StreamStatus_ENABLED :: StreamStatus

pattern StreamStatus_ENABLING :: StreamStatus

StreamViewType

newtype StreamViewType

Constructors

StreamViewType'

Fields

fromStreamViewType :: Text

Bundled Patterns

pattern StreamViewType_KEYS_ONLY :: StreamViewType

pattern StreamViewType_NEW_AND_OLD_IMAGES :: StreamViewType

pattern StreamViewType_NEW_IMAGE :: StreamViewType

pattern StreamViewType_OLD_IMAGE :: StreamViewType

AttributeValue

data AttributeValue

Represents the data for an attribute.

Each attribute value is described as a name-value pair. The name is the data type, and the value is the data itself.

For more information, see Data Types in the Amazon DynamoDB Developer Guide.

See: newAttributeValue smart constructor.

Constructors

AttributeValue' (Maybe [AttributeValue]) (Maybe [Text]) (Maybe (HashMap Text AttributeValue)) (Maybe Bool) (Maybe Text) (Maybe [Base64]) (Maybe Base64) (Maybe [Text]) (Maybe Text) (Maybe Bool)

newAttributeValue :: AttributeValue

Create a value of AttributeValue with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:l:AttributeValue', attributeValue_l - An attribute of type List. For example:

"L": [ {"S": "Cookies"} , {"S": "Coffee"}, {"N", "3.14159"}]

$sel:ns:AttributeValue', attributeValue_ns - An attribute of type Number Set. For example:

"NS": ["42.2", "-19", "7.5", "3.14"]

Numbers are sent across the network to DynamoDB as strings, to maximize compatibility across languages and libraries. However, DynamoDB treats them as number type attributes for mathematical operations.

$sel:m:AttributeValue', attributeValue_m - An attribute of type Map. For example:

"M": {"Name": {"S": "Joe"}, "Age": {"N": "35"}}

$sel:null:AttributeValue', attributeValue_null - An attribute of type Null. For example:

"NULL": true

$sel:n:AttributeValue', attributeValue_n - An attribute of type Number. For example:

"N": "123.45"

Numbers are sent across the network to DynamoDB as strings, to maximize compatibility across languages and libraries. However, DynamoDB treats them as number type attributes for mathematical operations.

$sel:bs:AttributeValue', attributeValue_bs - An attribute of type Binary Set. For example:

"BS": ["U3Vubnk=", "UmFpbnk=", "U25vd3k="]

$sel:b:AttributeValue', attributeValue_b - An attribute of type Binary. For example:

"B": "dGhpcyB0ZXh0IGlzIGJhc2U2NC1lbmNvZGVk"-- -- Note: This Lens automatically encodes and decodes Base64 data. -- The underlying isomorphism will encode to Base64 representation during -- serialisation, and decode from Base64 representation during deserialisation. -- This Lens accepts and returns only raw unencoded data.

$sel:ss:AttributeValue', attributeValue_ss - An attribute of type String Set. For example:

"SS": ["Giraffe", "Hippo" ,"Zebra"]

$sel:s:AttributeValue', attributeValue_s - An attribute of type String. For example:

"S": "Hello"

$sel:bool:AttributeValue', attributeValue_bool - An attribute of type Boolean. For example:

"BOOL": true

Identity

data Identity

Contains details about the type of identity that made the request.

See: newIdentity smart constructor.

Constructors

Identity' (Maybe Text) (Maybe Text)

newIdentity :: Identity

Create a value of Identity with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:principalId:Identity', identity_principalId - A unique identifier for the entity that made the call. For Time To Live, the principalId is "dynamodb.amazonaws.com".

$sel:type':Identity', identity_type - The type of the identity. For Time To Live, the type is "Service".

KeySchemaElement

data KeySchemaElement

Represents a single element of a key schema. A key schema specifies the attributes that make up the primary key of a table, or the key attributes of an index.

A KeySchemaElement represents exactly one attribute of the primary key. For example, a simple primary key would be represented by one KeySchemaElement (for the partition key). A composite primary key would require one KeySchemaElement for the partition key, and another KeySchemaElement for the sort key.

A KeySchemaElement must be a scalar, top-level attribute (not a nested attribute). The data type must be one of String, Number, or Binary. The attribute cannot be nested within a List or a Map.

See: newKeySchemaElement smart constructor.

Constructors

KeySchemaElement' Text KeyType

newKeySchemaElement

Arguments

:: Text
$sel:attributeName:KeySchemaElement'

-> KeyType
$sel:keyType:KeySchemaElement'

-> KeySchemaElement

Create a value of KeySchemaElement with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:attributeName:KeySchemaElement', keySchemaElement_attributeName - The name of a key attribute.

$sel:keyType:KeySchemaElement', keySchemaElement_keyType - The role that this key attribute will assume:

HASH - partition key

RANGE - sort key

The partition key of an item is also known as its hash attribute. The term "hash attribute" derives from DynamoDB's usage of an internal hash function to evenly distribute data items across partitions, based on their partition key values.

The sort key of an item is also known as its range attribute. The term "range attribute" derives from the way DynamoDB stores items with the same partition key physically close together, in sorted order by the sort key value.

Record

data Record

A description of a unique event within a stream.

See: newRecord smart constructor.

Constructors

Record' (Maybe Identity) (Maybe Text) (Maybe StreamRecord) (Maybe Text) (Maybe OperationType) (Maybe Text) (Maybe Text)

newRecord :: Record

Create a value of Record with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:userIdentity:Record', record_userIdentity - Items that are deleted by the Time to Live process after expiration have the following fields:

Records[].userIdentity.type

"Service"

Records[].userIdentity.principalId

"dynamodb.amazonaws.com"

$sel:eventVersion:Record', record_eventVersion - The version number of the stream record format. This number is updated whenever the structure of Record is modified.

Client applications must not assume that eventVersion will remain at a particular value, as this number is subject to change at any time. In general, eventVersion will only increase as the low-level DynamoDB Streams API evolves.

$sel:dynamodb:Record', record_dynamodb - The main body of the stream record, containing all of the DynamoDB-specific fields.

$sel:awsRegion:Record', record_awsRegion - The region in which the GetRecords request was received.

$sel:eventName:Record', record_eventName - The type of data modification that was performed on the DynamoDB table:

INSERT - a new item was added to the table.

MODIFY - one or more of an existing item's attributes were modified.

REMOVE - the item was deleted from the table

$sel:eventSource:Record', record_eventSource - The AWS service from which the stream record originated. For DynamoDB Streams, this is aws:dynamodb.

$sel:eventID:Record', record_eventID - A globally unique identifier for the event that was recorded in this stream record.

SequenceNumberRange

data SequenceNumberRange

The beginning and ending sequence numbers for the stream records contained within a shard.

See: newSequenceNumberRange smart constructor.

Constructors

SequenceNumberRange' (Maybe Text) (Maybe Text)

newSequenceNumberRange :: SequenceNumberRange

Create a value of SequenceNumberRange with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:startingSequenceNumber:SequenceNumberRange', sequenceNumberRange_startingSequenceNumber - The first sequence number for the stream records contained within a shard. String contains numeric characters only.

$sel:endingSequenceNumber:SequenceNumberRange', sequenceNumberRange_endingSequenceNumber - The last sequence number for the stream records contained within a shard. String contains numeric characters only.

Shard

data Shard

A uniquely identified group of stream records within a stream.

See: newShard smart constructor.

Constructors

Shard' (Maybe Text) (Maybe SequenceNumberRange) (Maybe Text)

newShard :: Shard

Create a value of Shard with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:parentShardId:Shard', shard_parentShardId - The shard ID of the current shard's parent.

$sel:sequenceNumberRange:Shard', shard_sequenceNumberRange - The range of possible sequence numbers for the shard.

$sel:shardId:Shard', shard_shardId - The system-generated identifier for this shard.

Stream

data Stream

Represents all of the data describing a particular stream.

See: newStream smart constructor.

Constructors

Stream' (Maybe Text) (Maybe Text) (Maybe Text)

newStream :: Stream

Create a value of Stream with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:streamLabel:Stream', stream_streamLabel - A timestamp, in ISO 8601 format, for this stream.

Note that LatestStreamLabel is not a unique identifier for the stream, because it is possible that a stream from another table might have the same timestamp. However, the combination of the following three elements is guaranteed to be unique:

the AWS customer ID.

the table name

the StreamLabel

$sel:streamArn:Stream', stream_streamArn - The Amazon Resource Name (ARN) for the stream.

$sel:tableName:Stream', stream_tableName - The DynamoDB table with which the stream is associated.

StreamDescription

data StreamDescription

Represents all of the data describing a particular stream.

See: newStreamDescription smart constructor.

Constructors

StreamDescription' (Maybe Text) (Maybe Text) (Maybe StreamStatus) (Maybe (NonEmpty KeySchemaElement)) (Maybe StreamViewType) (Maybe Text) (Maybe [Shard]) (Maybe Text) (Maybe POSIX)

newStreamDescription :: StreamDescription

Create a value of StreamDescription with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:lastEvaluatedShardId:StreamDescription', streamDescription_lastEvaluatedShardId - The shard ID of the item where the operation stopped, inclusive of the previous result set. Use this value to start a new operation, excluding this value in the new request.

If LastEvaluatedShardId is empty, then the "last page" of results has been processed and there is currently no more data to be retrieved.

If LastEvaluatedShardId is not empty, it does not necessarily mean that there is more data in the result set. The only way to know when you have reached the end of the result set is when LastEvaluatedShardId is empty.

$sel:streamLabel:StreamDescription', streamDescription_streamLabel - A timestamp, in ISO 8601 format, for this stream.

Note that LatestStreamLabel is not a unique identifier for the stream, because it is possible that a stream from another table might have the same timestamp. However, the combination of the following three elements is guaranteed to be unique:

the AWS customer ID.

the table name

the StreamLabel

$sel:streamStatus:StreamDescription', streamDescription_streamStatus - Indicates the current status of the stream:

ENABLING - Streams is currently being enabled on the DynamoDB table.

ENABLED - the stream is enabled.

DISABLING - Streams is currently being disabled on the DynamoDB table.

DISABLED - the stream is disabled.

$sel:keySchema:StreamDescription', streamDescription_keySchema - The key attribute(s) of the stream's DynamoDB table.

$sel:streamViewType:StreamDescription', streamDescription_streamViewType - Indicates the format of the records within this stream:

KEYS_ONLY - only the key attributes of items that were modified in the DynamoDB table.

NEW_IMAGE - entire items from the table, as they appeared after they were modified.

OLD_IMAGE - entire items from the table, as they appeared before they were modified.

NEW_AND_OLD_IMAGES - both the new and the old images of the items from the table.

$sel:streamArn:StreamDescription', streamDescription_streamArn - The Amazon Resource Name (ARN) for the stream.

$sel:shards:StreamDescription', streamDescription_shards - The shards that comprise the stream.

$sel:tableName:StreamDescription', streamDescription_tableName - The DynamoDB table with which the stream is associated.

$sel:creationRequestDateTime:StreamDescription', streamDescription_creationRequestDateTime - The date and time when the request to create this stream was issued.

StreamRecord

data StreamRecord

A description of a single data modification that was performed on an item in a DynamoDB table.

See: newStreamRecord smart constructor.

Constructors

StreamRecord' (Maybe Natural) (Maybe Text) (Maybe POSIX) (Maybe StreamViewType) (Maybe (HashMap Text AttributeValue)) (Maybe (HashMap Text AttributeValue)) (Maybe (HashMap Text AttributeValue))

newStreamRecord :: StreamRecord

Create a value of StreamRecord with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:sizeBytes:StreamRecord', streamRecord_sizeBytes - The size of the stream record, in bytes.

$sel:sequenceNumber:StreamRecord', streamRecord_sequenceNumber - The sequence number of the stream record.

$sel:approximateCreationDateTime:StreamRecord', streamRecord_approximateCreationDateTime - The approximate date and time when the stream record was created, in UNIX epoch time format.

$sel:streamViewType:StreamRecord', streamRecord_streamViewType - The type of data from the modified DynamoDB item that was captured in this stream record:

KEYS_ONLY - only the key attributes of the modified item.

NEW_IMAGE - the entire item, as it appeared after it was modified.

OLD_IMAGE - the entire item, as it appeared before it was modified.

NEW_AND_OLD_IMAGES - both the new and the old item images of the item.

$sel:keys:StreamRecord', streamRecord_keys - The primary key attribute(s) for the DynamoDB item that was modified.

$sel:oldImage:StreamRecord', streamRecord_oldImage - The item in the DynamoDB table as it appeared before it was modified.

$sel:newImage':StreamRecord', streamRecord_newImage - The item in the DynamoDB table as it appeared after it was modified.