Amazonka.DynamoDBStreams.Types.StreamRecord

Description

Documentation

data StreamRecord

A description of a single data modification that was performed on an item in a DynamoDB table.

See: newStreamRecord smart constructor.

Constructors

StreamRecord'

Fields

sizeBytes :: Maybe Natural

The size of the stream record, in bytes.

sequenceNumber :: Maybe Text

The sequence number of the stream record.

approximateCreationDateTime :: Maybe POSIX

The approximate date and time when the stream record was created, in UNIX epoch time format.

streamViewType :: Maybe StreamViewType

The type of data from the modified DynamoDB item that was captured in this stream record:

KEYS_ONLY - only the key attributes of the modified item.

NEW_IMAGE - the entire item, as it appeared after it was modified.

OLD_IMAGE - the entire item, as it appeared before it was modified.

NEW_AND_OLD_IMAGES - both the new and the old item images of the item.

keys :: Maybe (HashMap Text AttributeValue)

The primary key attribute(s) for the DynamoDB item that was modified.

oldImage :: Maybe (HashMap Text AttributeValue)

The item in the DynamoDB table as it appeared before it was modified.

newImage' :: Maybe (HashMap Text AttributeValue)

The item in the DynamoDB table as it appeared after it was modified.

newStreamRecord :: StreamRecord

Create a value of StreamRecord with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:sizeBytes:StreamRecord', streamRecord_sizeBytes - The size of the stream record, in bytes.

$sel:sequenceNumber:StreamRecord', streamRecord_sequenceNumber - The sequence number of the stream record.

$sel:approximateCreationDateTime:StreamRecord', streamRecord_approximateCreationDateTime - The approximate date and time when the stream record was created, in UNIX epoch time format.

$sel:streamViewType:StreamRecord', streamRecord_streamViewType - The type of data from the modified DynamoDB item that was captured in this stream record:

KEYS_ONLY - only the key attributes of the modified item.

NEW_IMAGE - the entire item, as it appeared after it was modified.

OLD_IMAGE - the entire item, as it appeared before it was modified.

NEW_AND_OLD_IMAGES - both the new and the old item images of the item.

$sel:keys:StreamRecord', streamRecord_keys - The primary key attribute(s) for the DynamoDB item that was modified.

$sel:oldImage:StreamRecord', streamRecord_oldImage - The item in the DynamoDB table as it appeared before it was modified.

$sel:newImage':StreamRecord', streamRecord_newImage - The item in the DynamoDB table as it appeared after it was modified.

streamRecord_sizeBytes :: Lens' StreamRecord (Maybe Natural)

The size of the stream record, in bytes.

streamRecord_sequenceNumber :: Lens' StreamRecord (Maybe Text)

The sequence number of the stream record.

streamRecord_approximateCreationDateTime :: Lens' StreamRecord (Maybe UTCTime)

The approximate date and time when the stream record was created, in UNIX epoch time format.

streamRecord_streamViewType :: Lens' StreamRecord (Maybe StreamViewType)

The type of data from the modified DynamoDB item that was captured in this stream record:

KEYS_ONLY - only the key attributes of the modified item.

NEW_IMAGE - the entire item, as it appeared after it was modified.

OLD_IMAGE - the entire item, as it appeared before it was modified.

NEW_AND_OLD_IMAGES - both the new and the old item images of the item.

streamRecord_keys :: Lens' StreamRecord (Maybe (HashMap Text AttributeValue))

The primary key attribute(s) for the DynamoDB item that was modified.

streamRecord_oldImage :: Lens' StreamRecord (Maybe (HashMap Text AttributeValue))

The item in the DynamoDB table as it appeared before it was modified.

streamRecord_newImage :: Lens' StreamRecord (Maybe (HashMap Text AttributeValue))

The item in the DynamoDB table as it appeared after it was modified.