Amazonka.EBS

Description

Derived from API version 2019-11-02 of the AWS service descriptions, licensed under Apache 2.0.

You can use the Amazon Elastic Block Store (Amazon EBS) direct APIs to create Amazon EBS snapshots, write data directly to your snapshots, read data on your snapshots, and identify the differences or changes between two snapshots. If you’re an independent software vendor (ISV) who offers backup services for Amazon EBS, the EBS direct APIs make it more efficient and cost-effective to track incremental changes on your Amazon EBS volumes through snapshots. This can be done without having to create new volumes from snapshots, and then use Amazon Elastic Compute Cloud (Amazon EC2) instances to compare the differences.

You can create incremental snapshots directly from data on-premises into volumes and the cloud to use for quick disaster recovery. With the ability to write and read snapshots, you can write your on-premises data to an snapshot during a disaster. Then after recovery, you can restore it back to Amazon Web Services or on-premises from the snapshot. You no longer need to build and maintain complex mechanisms to copy data to and from Amazon EBS.

This API reference provides detailed information about the actions, data types, parameters, and errors of the EBS direct APIs. For more information about the elements that make up the EBS direct APIs, and examples of how to use them effectively, see Accessing the Contents of an Amazon EBS Snapshot in the Amazon Elastic Compute Cloud User Guide. For more information about the supported Amazon Web Services Regions, endpoints, and service quotas for the EBS direct APIs, see Amazon Elastic Block Store Endpoints and Quotas in the Amazon Web Services General Reference.

Service Configuration

defaultService :: Service

API version 2019-11-02 of the Amazon Elastic Block Store SDK configuration.

Errors

Error matchers are designed for use with the functions provided by Control.Exception.Lens. This allows catching (and rethrowing) service specific errors returned by EBS.

ValidationException

_ValidationException :: AsError a => Getting (First ServiceError) a ServiceError

The input fails to satisfy the constraints of the EBS direct APIs.

AccessDeniedException

_AccessDeniedException :: AsError a => Getting (First ServiceError) a ServiceError

You do not have sufficient access to perform this action.

ConflictException

_ConflictException :: AsError a => Getting (First ServiceError) a ServiceError

The request uses the same client token as a previous, but non-identical request.

ServiceQuotaExceededException

_ServiceQuotaExceededException :: AsError a => Getting (First ServiceError) a ServiceError

Your current service quotas do not allow you to perform this action.

ConcurrentLimitExceededException

_ConcurrentLimitExceededException :: AsError a => Getting (First ServiceError) a ServiceError

You have reached the limit for concurrent API requests. For more information, see Optimizing performance of the EBS direct APIs in the Amazon Elastic Compute Cloud User Guide.

RequestThrottledException

_RequestThrottledException :: AsError a => Getting (First ServiceError) a ServiceError

The number of API requests has exceed the maximum allowed API request throttling limit.

InternalServerException

_InternalServerException :: AsError a => Getting (First ServiceError) a ServiceError

An internal error has occurred.

ResourceNotFoundException

_ResourceNotFoundException :: AsError a => Getting (First ServiceError) a ServiceError

The specified resource does not exist.

Waiters

Waiters poll by repeatedly sending a request until some remote success condition configured by the Wait specification is fulfilled. The Wait specification determines how many attempts should be made, in addition to delay and retry strategies.

Operations

Some AWS operations return results that are incomplete and require subsequent requests in order to obtain the entire result set. The process of sending subsequent requests to continue where a previous request left off is called pagination. For example, the ListObjects operation of Amazon S3 returns up to 1000 objects at a time, and you must send subsequent requests with the appropriate Marker in order to retrieve the next page of results.

Operations that have an AWSPager instance can transparently perform subsequent requests, correctly setting Markers and other request facets to iterate through the entire result set of a truncated API operation. Operations which support this have an additional note in the documentation.

Many operations have the ability to filter results on the server side. See the individual operation parameters for details.

StartSnapshot

data StartSnapshot

See: newStartSnapshot smart constructor.

Constructors

StartSnapshot' (Maybe (Sensitive Text)) (Maybe Text) (Maybe Bool) (Maybe Text) (Maybe Natural) (Maybe Text) (Maybe [Tag]) Natural

newStartSnapshot

Arguments

:: Natural
$sel:volumeSize:StartSnapshot'

-> StartSnapshot

Create a value of StartSnapshot with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:kmsKeyArn:StartSnapshot', startSnapshot_kmsKeyArn - The Amazon Resource Name (ARN) of the Key Management Service (KMS) key to be used to encrypt the snapshot. If you do not specify a KMS key, the default Amazon Web Services managed KMS key is used.

If you specify a ParentSnapshotId, omit this parameter; the snapshot will be encrypted using the same KMS key that was used to encrypt the parent snapshot.

If Encrypted is set to true, you must specify a KMS key ARN.

$sel:clientToken:StartSnapshot', startSnapshot_clientToken - A unique, case-sensitive identifier that you provide to ensure the idempotency of the request. Idempotency ensures that an API request completes only once. With an idempotent request, if the original request completes successfully. The subsequent retries with the same client token return the result from the original successful request and they have no additional effect.

If you do not specify a client token, one is automatically generated by the Amazon Web Services SDK.

For more information, see Idempotency for StartSnapshot API in the Amazon Elastic Compute Cloud User Guide.

$sel:encrypted:StartSnapshot', startSnapshot_encrypted - Indicates whether to encrypt the snapshot. To create an encrypted snapshot, specify true. To create an unencrypted snapshot, omit this parameter.

If you specify a value for ParentSnapshotId, omit this parameter.

If you specify true, the snapshot is encrypted using the KMS key specified using the KmsKeyArn parameter. If no value is specified for KmsKeyArn, the default KMS key for your account is used. If no default KMS key has been specified for your account, the Amazon Web Services managed KMS key is used. To set a default KMS key for your account, use ModifyEbsDefaultKmsKeyId.

If your account is enabled for encryption by default, you cannot set this parameter to false. In this case, you can omit this parameter.

For more information, see Using encryption in the Amazon Elastic Compute Cloud User Guide.

$sel:parentSnapshotId:StartSnapshot', startSnapshot_parentSnapshotId - The ID of the parent snapshot. If there is no parent snapshot, or if you are creating the first snapshot for an on-premises volume, omit this parameter.

If your account is enabled for encryption by default, you cannot use an unencrypted snapshot as a parent snapshot. You must first create an encrypted copy of the parent snapshot using CopySnapshot.

$sel:timeout:StartSnapshot', startSnapshot_timeout - The amount of time (in minutes) after which the snapshot is automatically cancelled if:

No blocks are written to the snapshot.

The snapshot is not completed after writing the last block of data.

If no value is specified, the timeout defaults to 60 minutes.

$sel:description:StartSnapshot', startSnapshot_description - A description for the snapshot.

$sel:tags:StartSnapshot', startSnapshot_tags - The tags to apply to the snapshot.

$sel:volumeSize:StartSnapshot', startSnapshot_volumeSize - The size of the volume, in GiB. The maximum size is 65536 GiB (64 TiB).

data StartSnapshotResponse

See: newStartSnapshotResponse smart constructor.

Constructors

StartSnapshotResponse' (Maybe Int) (Maybe Status) (Maybe (Sensitive Text)) (Maybe POSIX) (Maybe Natural) (Maybe Text) (Maybe Text) (Maybe Text) (Maybe [Tag]) (Maybe Text) Int

newStartSnapshotResponse

Arguments

:: Int
$sel:httpStatus:StartSnapshotResponse'

-> StartSnapshotResponse

Create a value of StartSnapshotResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:blockSize:StartSnapshotResponse', startSnapshotResponse_blockSize - The size of the blocks in the snapshot, in bytes.

$sel:status:StartSnapshotResponse', startSnapshotResponse_status - The status of the snapshot.

$sel:kmsKeyArn:StartSnapshot', startSnapshotResponse_kmsKeyArn - The Amazon Resource Name (ARN) of the Key Management Service (KMS) key used to encrypt the snapshot.

$sel:startTime:StartSnapshotResponse', startSnapshotResponse_startTime - The timestamp when the snapshot was created.

$sel:volumeSize:StartSnapshot', startSnapshotResponse_volumeSize - The size of the volume, in GiB.

$sel:ownerId:StartSnapshotResponse', startSnapshotResponse_ownerId - The Amazon Web Services account ID of the snapshot owner.

$sel:parentSnapshotId:StartSnapshot', startSnapshotResponse_parentSnapshotId - The ID of the parent snapshot.

$sel:description:StartSnapshot', startSnapshotResponse_description - The description of the snapshot.

$sel:tags:StartSnapshot', startSnapshotResponse_tags - The tags applied to the snapshot. You can specify up to 50 tags per snapshot. For more information, see Tagging your Amazon EC2 resources in the Amazon Elastic Compute Cloud User Guide.

$sel:snapshotId:StartSnapshotResponse', startSnapshotResponse_snapshotId - The ID of the snapshot.

$sel:httpStatus:StartSnapshotResponse', startSnapshotResponse_httpStatus - The response's http status code.

ListSnapshotBlocks

data ListSnapshotBlocks

See: newListSnapshotBlocks smart constructor.

Constructors

ListSnapshotBlocks' (Maybe Text) (Maybe Natural) (Maybe Natural) Text

newListSnapshotBlocks

Arguments

:: Text
$sel:snapshotId:ListSnapshotBlocks'

-> ListSnapshotBlocks

Create a value of ListSnapshotBlocks with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:nextToken:ListSnapshotBlocks', listSnapshotBlocks_nextToken - The token to request the next page of results.

$sel:startingBlockIndex:ListSnapshotBlocks', listSnapshotBlocks_startingBlockIndex - The block index from which the list should start. The list in the response will start from this block index or the next valid block index in the snapshot.

$sel:maxResults:ListSnapshotBlocks', listSnapshotBlocks_maxResults - The number of results to return.

$sel:snapshotId:ListSnapshotBlocks', listSnapshotBlocks_snapshotId - The ID of the snapshot from which to get block indexes and block tokens.

data ListSnapshotBlocksResponse

See: newListSnapshotBlocksResponse smart constructor.

Constructors

ListSnapshotBlocksResponse' (Maybe Int) (Maybe (Sensitive [Block])) (Maybe POSIX) (Maybe Natural) (Maybe Text) Int

newListSnapshotBlocksResponse

Arguments

:: Int
$sel:httpStatus:ListSnapshotBlocksResponse'

-> ListSnapshotBlocksResponse

Create a value of ListSnapshotBlocksResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:blockSize:ListSnapshotBlocksResponse', listSnapshotBlocksResponse_blockSize - The size of the blocks in the snapshot, in bytes.

$sel:blocks:ListSnapshotBlocksResponse', listSnapshotBlocksResponse_blocks - An array of objects containing information about the blocks.

$sel:expiryTime:ListSnapshotBlocksResponse', listSnapshotBlocksResponse_expiryTime - The time when the BlockToken expires.

$sel:volumeSize:ListSnapshotBlocksResponse', listSnapshotBlocksResponse_volumeSize - The size of the volume in GB.

$sel:nextToken:ListSnapshotBlocks', listSnapshotBlocksResponse_nextToken - The token to use to retrieve the next page of results. This value is null when there are no more results to return.

$sel:httpStatus:ListSnapshotBlocksResponse', listSnapshotBlocksResponse_httpStatus - The response's http status code.

PutSnapshotBlock

data PutSnapshotBlock

See: newPutSnapshotBlock smart constructor.

Constructors

PutSnapshotBlock' (Maybe Natural) Text Natural Int Text ChecksumAlgorithm HashedBody

newPutSnapshotBlock

Arguments

:: Text
$sel:snapshotId:PutSnapshotBlock'

-> Natural
$sel:blockIndex:PutSnapshotBlock'

-> Int
$sel:dataLength:PutSnapshotBlock'

-> Text
$sel:checksum:PutSnapshotBlock'

-> ChecksumAlgorithm
$sel:checksumAlgorithm:PutSnapshotBlock'

-> HashedBody
$sel:blockData:PutSnapshotBlock'

-> PutSnapshotBlock

Create a value of PutSnapshotBlock with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:progress:PutSnapshotBlock', putSnapshotBlock_progress - The progress of the write process, as a percentage.

$sel:snapshotId:PutSnapshotBlock', putSnapshotBlock_snapshotId - The ID of the snapshot.

$sel:blockIndex:PutSnapshotBlock', putSnapshotBlock_blockIndex - The block index of the block in which to write the data. A block index is a logical index in units of 512 KiB blocks. To identify the block index, divide the logical offset of the data in the logical volume by the block size (logical offset of data/524288). The logical offset of the data must be 512 KiB aligned.

$sel:dataLength:PutSnapshotBlock', putSnapshotBlock_dataLength - The size of the data to write to the block, in bytes. Currently, the only supported size is 524288 bytes.

Valid values: 524288

$sel:checksum:PutSnapshotBlock', putSnapshotBlock_checksum - A Base64-encoded SHA256 checksum of the data. Only SHA256 checksums are supported.

$sel:checksumAlgorithm:PutSnapshotBlock', putSnapshotBlock_checksumAlgorithm - The algorithm used to generate the checksum. Currently, the only supported algorithm is SHA256.

$sel:blockData:PutSnapshotBlock', putSnapshotBlock_blockData - The data to write to the block.

The block data is not signed as part of the Signature Version 4 signing process. As a result, you must generate and provide a Base64-encoded SHA256 checksum for the block data using the x-amz-Checksum header. Also, you must specify the checksum algorithm using the x-amz-Checksum-Algorithm header. The checksum that you provide is part of the Signature Version 4 signing process. It is validated against a checksum generated by Amazon EBS to ensure the validity and authenticity of the data. If the checksums do not correspond, the request fails. For more information, see Using checksums with the EBS direct APIs in the Amazon Elastic Compute Cloud User Guide.

data PutSnapshotBlockResponse

See: newPutSnapshotBlockResponse smart constructor.

Constructors

PutSnapshotBlockResponse' (Maybe ChecksumAlgorithm) (Maybe Text) Int

newPutSnapshotBlockResponse

Arguments

:: Int
$sel:httpStatus:PutSnapshotBlockResponse'

-> PutSnapshotBlockResponse

Create a value of PutSnapshotBlockResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:checksumAlgorithm:PutSnapshotBlock', putSnapshotBlockResponse_checksumAlgorithm - The algorithm used by Amazon EBS to generate the checksum.

$sel:checksum:PutSnapshotBlock', putSnapshotBlockResponse_checksum - The SHA256 checksum generated for the block data by Amazon EBS.

$sel:httpStatus:PutSnapshotBlockResponse', putSnapshotBlockResponse_httpStatus - The response's http status code.

ListChangedBlocks

data ListChangedBlocks

See: newListChangedBlocks smart constructor.

Constructors

ListChangedBlocks' (Maybe Text) (Maybe Natural) (Maybe Text) (Maybe Natural) Text

newListChangedBlocks

Arguments

:: Text
$sel:secondSnapshotId:ListChangedBlocks'

-> ListChangedBlocks

Create a value of ListChangedBlocks with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:nextToken:ListChangedBlocks', listChangedBlocks_nextToken - The token to request the next page of results.

$sel:startingBlockIndex:ListChangedBlocks', listChangedBlocks_startingBlockIndex - The block index from which the comparison should start.

The list in the response will start from this block index or the next valid block index in the snapshots.

$sel:firstSnapshotId:ListChangedBlocks', listChangedBlocks_firstSnapshotId - The ID of the first snapshot to use for the comparison.

The FirstSnapshotID parameter must be specified with a SecondSnapshotId parameter; otherwise, an error occurs.

$sel:maxResults:ListChangedBlocks', listChangedBlocks_maxResults - The number of results to return.

$sel:secondSnapshotId:ListChangedBlocks', listChangedBlocks_secondSnapshotId - The ID of the second snapshot to use for the comparison.

The SecondSnapshotId parameter must be specified with a FirstSnapshotID parameter; otherwise, an error occurs.

data ListChangedBlocksResponse

See: newListChangedBlocksResponse smart constructor.

Constructors

ListChangedBlocksResponse' (Maybe Int) (Maybe POSIX) (Maybe Natural) (Maybe [Sensitive ChangedBlock]) (Maybe Text) Int

newListChangedBlocksResponse

Arguments

:: Int
$sel:httpStatus:ListChangedBlocksResponse'

-> ListChangedBlocksResponse

Create a value of ListChangedBlocksResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:blockSize:ListChangedBlocksResponse', listChangedBlocksResponse_blockSize - The size of the blocks in the snapshot, in bytes.

$sel:expiryTime:ListChangedBlocksResponse', listChangedBlocksResponse_expiryTime - The time when the BlockToken expires.

$sel:volumeSize:ListChangedBlocksResponse', listChangedBlocksResponse_volumeSize - The size of the volume in GB.

$sel:changedBlocks:ListChangedBlocksResponse', listChangedBlocksResponse_changedBlocks - An array of objects containing information about the changed blocks.

$sel:nextToken:ListChangedBlocks', listChangedBlocksResponse_nextToken - The token to use to retrieve the next page of results. This value is null when there are no more results to return.

$sel:httpStatus:ListChangedBlocksResponse', listChangedBlocksResponse_httpStatus - The response's http status code.

CompleteSnapshot

data CompleteSnapshot

See: newCompleteSnapshot smart constructor.

Constructors

CompleteSnapshot' (Maybe ChecksumAlgorithm) (Maybe Text) (Maybe ChecksumAggregationMethod) Text Natural

newCompleteSnapshot

Arguments

:: Text
$sel:snapshotId:CompleteSnapshot'

-> Natural
$sel:changedBlocksCount:CompleteSnapshot'

-> CompleteSnapshot

Create a value of CompleteSnapshot with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:checksumAlgorithm:CompleteSnapshot', completeSnapshot_checksumAlgorithm - The algorithm used to generate the checksum. Currently, the only supported algorithm is SHA256.

$sel:checksum:CompleteSnapshot', completeSnapshot_checksum - An aggregated Base-64 SHA256 checksum based on the checksums of each written block.

To generate the aggregated checksum using the linear aggregation method, arrange the checksums for each written block in ascending order of their block index, concatenate them to form a single string, and then generate the checksum on the entire string using the SHA256 algorithm.

$sel:checksumAggregationMethod:CompleteSnapshot', completeSnapshot_checksumAggregationMethod - The aggregation method used to generate the checksum. Currently, the only supported aggregation method is LINEAR.

$sel:snapshotId:CompleteSnapshot', completeSnapshot_snapshotId - The ID of the snapshot.

$sel:changedBlocksCount:CompleteSnapshot', completeSnapshot_changedBlocksCount - The number of blocks that were written to the snapshot.

data CompleteSnapshotResponse

See: newCompleteSnapshotResponse smart constructor.

Constructors

CompleteSnapshotResponse' (Maybe Status) Int

newCompleteSnapshotResponse

Arguments

:: Int
$sel:httpStatus:CompleteSnapshotResponse'

-> CompleteSnapshotResponse

Create a value of CompleteSnapshotResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:status:CompleteSnapshotResponse', completeSnapshotResponse_status - The status of the snapshot.

$sel:httpStatus:CompleteSnapshotResponse', completeSnapshotResponse_httpStatus - The response's http status code.

GetSnapshotBlock

data GetSnapshotBlock

See: newGetSnapshotBlock smart constructor.

Constructors

GetSnapshotBlock' Text Natural Text

newGetSnapshotBlock

Arguments

:: Text
$sel:snapshotId:GetSnapshotBlock'

-> Natural
$sel:blockIndex:GetSnapshotBlock'

-> Text
$sel:blockToken:GetSnapshotBlock'

-> GetSnapshotBlock

Create a value of GetSnapshotBlock with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:snapshotId:GetSnapshotBlock', getSnapshotBlock_snapshotId - The ID of the snapshot containing the block from which to get data.

$sel:blockIndex:GetSnapshotBlock', getSnapshotBlock_blockIndex - The block index of the block from which to get data.

Obtain the BlockIndex by running the ListChangedBlocks or ListSnapshotBlocks operations.

$sel:blockToken:GetSnapshotBlock', getSnapshotBlock_blockToken - The block token of the block from which to get data.

Obtain the BlockToken by running the ListChangedBlocks or ListSnapshotBlocks operations.

data GetSnapshotBlockResponse

See: newGetSnapshotBlockResponse smart constructor.

Constructors

GetSnapshotBlockResponse' (Maybe Int) (Maybe ChecksumAlgorithm) (Maybe Text) Int ResponseBody

newGetSnapshotBlockResponse

Arguments

:: Int
$sel:httpStatus:GetSnapshotBlockResponse'

-> ResponseBody
$sel:blockData:GetSnapshotBlockResponse'

-> GetSnapshotBlockResponse

Create a value of GetSnapshotBlockResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:dataLength:GetSnapshotBlockResponse', getSnapshotBlockResponse_dataLength - The size of the data in the block.

$sel:checksumAlgorithm:GetSnapshotBlockResponse', getSnapshotBlockResponse_checksumAlgorithm - The algorithm used to generate the checksum for the block, such as SHA256.

$sel:checksum:GetSnapshotBlockResponse', getSnapshotBlockResponse_checksum - The checksum generated for the block, which is Base64 encoded.

$sel:httpStatus:GetSnapshotBlockResponse', getSnapshotBlockResponse_httpStatus - The response's http status code.

$sel:blockData:GetSnapshotBlockResponse', getSnapshotBlockResponse_blockData - The data content of the block.

Types

ChecksumAggregationMethod

newtype ChecksumAggregationMethod

Constructors

ChecksumAggregationMethod'

Fields

fromChecksumAggregationMethod :: Text

Bundled Patterns

pattern ChecksumAggregationMethod_LINEAR :: ChecksumAggregationMethod

ChecksumAlgorithm

newtype ChecksumAlgorithm

Constructors

ChecksumAlgorithm'

Fields

fromChecksumAlgorithm :: Text

Bundled Patterns

pattern ChecksumAlgorithm_SHA256 :: ChecksumAlgorithm

Status

newtype Status

Constructors

Status'

Fields

fromStatus :: Text

Bundled Patterns

pattern Status_Completed :: Status

pattern Status_Error :: Status

pattern Status_Pending :: Status

Block

data Block

A block of data in an Amazon Elastic Block Store snapshot.

See: newBlock smart constructor.

Constructors

Block' (Maybe Natural) (Maybe Text)

newBlock :: Block

Create a value of Block with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:blockIndex:Block', block_blockIndex - The block index.

$sel:blockToken:Block', block_blockToken - The block token for the block index.

ChangedBlock

data ChangedBlock

A block of data in an Amazon Elastic Block Store snapshot that is different from another snapshot of the same volume/snapshot lineage.

See: newChangedBlock smart constructor.

Constructors

ChangedBlock' (Maybe Natural) (Maybe Text) (Maybe Text)

newChangedBlock :: ChangedBlock

Create a value of ChangedBlock with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:blockIndex:ChangedBlock', changedBlock_blockIndex - The block index.

$sel:secondBlockToken:ChangedBlock', changedBlock_secondBlockToken - The block token for the block index of the SecondSnapshotId specified in the ListChangedBlocks operation.

$sel:firstBlockToken:ChangedBlock', changedBlock_firstBlockToken - The block token for the block index of the FirstSnapshotId specified in the ListChangedBlocks operation. This value is absent if the first snapshot does not have the changed block that is on the second snapshot.

Tag

data Tag

Describes a tag.

See: newTag smart constructor.

Constructors

Tag' (Maybe Text) (Maybe Text)

newTag :: Tag

Create a value of Tag with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:value:Tag', tag_value - The value of the tag.

$sel:key:Tag', tag_key - The key of the tag.