Amazonka.EBS.GetSnapshotBlock

Description

Returns the data in a block in an Amazon Elastic Block Store snapshot.

Creating a Request

data GetSnapshotBlock

See: newGetSnapshotBlock smart constructor.

Constructors

GetSnapshotBlock'

Fields

snapshotId :: Text

The ID of the snapshot containing the block from which to get data.

blockIndex :: Natural

The block index of the block from which to get data.

Obtain the BlockIndex by running the ListChangedBlocks or ListSnapshotBlocks operations.

blockToken :: Text

The block token of the block from which to get data.

Obtain the BlockToken by running the ListChangedBlocks or ListSnapshotBlocks operations.

newGetSnapshotBlock

Arguments

:: Text
$sel:snapshotId:GetSnapshotBlock'

-> Natural
$sel:blockIndex:GetSnapshotBlock'

-> Text
$sel:blockToken:GetSnapshotBlock'

-> GetSnapshotBlock

Create a value of GetSnapshotBlock with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:snapshotId:GetSnapshotBlock', getSnapshotBlock_snapshotId - The ID of the snapshot containing the block from which to get data.

$sel:blockIndex:GetSnapshotBlock', getSnapshotBlock_blockIndex - The block index of the block from which to get data.

Obtain the BlockIndex by running the ListChangedBlocks or ListSnapshotBlocks operations.

$sel:blockToken:GetSnapshotBlock', getSnapshotBlock_blockToken - The block token of the block from which to get data.

Obtain the BlockToken by running the ListChangedBlocks or ListSnapshotBlocks operations.

Request Lenses

getSnapshotBlock_snapshotId :: Lens' GetSnapshotBlock Text

The ID of the snapshot containing the block from which to get data.

getSnapshotBlock_blockIndex :: Lens' GetSnapshotBlock Natural

The block index of the block from which to get data.

Obtain the BlockIndex by running the ListChangedBlocks or ListSnapshotBlocks operations.

getSnapshotBlock_blockToken :: Lens' GetSnapshotBlock Text

The block token of the block from which to get data.

Obtain the BlockToken by running the ListChangedBlocks or ListSnapshotBlocks operations.

Destructuring the Response

data GetSnapshotBlockResponse

See: newGetSnapshotBlockResponse smart constructor.

Constructors

GetSnapshotBlockResponse'

Fields

dataLength :: Maybe Int

The size of the data in the block.

checksumAlgorithm :: Maybe ChecksumAlgorithm

The algorithm used to generate the checksum for the block, such as SHA256.

checksum :: Maybe Text

The checksum generated for the block, which is Base64 encoded.

httpStatus :: Int

The response's http status code.

blockData :: ResponseBody

The data content of the block.

newGetSnapshotBlockResponse

Arguments

:: Int
$sel:httpStatus:GetSnapshotBlockResponse'

-> ResponseBody
$sel:blockData:GetSnapshotBlockResponse'

-> GetSnapshotBlockResponse

Create a value of GetSnapshotBlockResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:dataLength:GetSnapshotBlockResponse', getSnapshotBlockResponse_dataLength - The size of the data in the block.

$sel:checksumAlgorithm:GetSnapshotBlockResponse', getSnapshotBlockResponse_checksumAlgorithm - The algorithm used to generate the checksum for the block, such as SHA256.

$sel:checksum:GetSnapshotBlockResponse', getSnapshotBlockResponse_checksum - The checksum generated for the block, which is Base64 encoded.

$sel:httpStatus:GetSnapshotBlockResponse', getSnapshotBlockResponse_httpStatus - The response's http status code.

$sel:blockData:GetSnapshotBlockResponse', getSnapshotBlockResponse_blockData - The data content of the block.

Response Lenses

getSnapshotBlockResponse_dataLength :: Lens' GetSnapshotBlockResponse (Maybe Int)

The size of the data in the block.

getSnapshotBlockResponse_checksumAlgorithm :: Lens' GetSnapshotBlockResponse (Maybe ChecksumAlgorithm)

The algorithm used to generate the checksum for the block, such as SHA256.

getSnapshotBlockResponse_checksum :: Lens' GetSnapshotBlockResponse (Maybe Text)

The checksum generated for the block, which is Base64 encoded.

getSnapshotBlockResponse_httpStatus :: Lens' GetSnapshotBlockResponse Int

The response's http status code.

getSnapshotBlockResponse_blockData :: Lens' GetSnapshotBlockResponse ResponseBody

The data content of the block.