Amazonka.EBS.ListChangedBlocks

Description

Returns information about the blocks that are different between two Amazon Elastic Block Store snapshots of the same volume/snapshot lineage.

Creating a Request

data ListChangedBlocks

See: newListChangedBlocks smart constructor.

Constructors

ListChangedBlocks'

Fields

nextToken :: Maybe Text

The token to request the next page of results.

startingBlockIndex :: Maybe Natural

The block index from which the comparison should start.

The list in the response will start from this block index or the next valid block index in the snapshots.

firstSnapshotId :: Maybe Text

The ID of the first snapshot to use for the comparison.

The FirstSnapshotID parameter must be specified with a SecondSnapshotId parameter; otherwise, an error occurs.

maxResults :: Maybe Natural

The number of results to return.

secondSnapshotId :: Text

The ID of the second snapshot to use for the comparison.

The SecondSnapshotId parameter must be specified with a FirstSnapshotID parameter; otherwise, an error occurs.

newListChangedBlocks

Arguments

:: Text
$sel:secondSnapshotId:ListChangedBlocks'

-> ListChangedBlocks

Create a value of ListChangedBlocks with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:nextToken:ListChangedBlocks', listChangedBlocks_nextToken - The token to request the next page of results.

$sel:startingBlockIndex:ListChangedBlocks', listChangedBlocks_startingBlockIndex - The block index from which the comparison should start.

The list in the response will start from this block index or the next valid block index in the snapshots.

$sel:firstSnapshotId:ListChangedBlocks', listChangedBlocks_firstSnapshotId - The ID of the first snapshot to use for the comparison.

The FirstSnapshotID parameter must be specified with a SecondSnapshotId parameter; otherwise, an error occurs.

$sel:maxResults:ListChangedBlocks', listChangedBlocks_maxResults - The number of results to return.

$sel:secondSnapshotId:ListChangedBlocks', listChangedBlocks_secondSnapshotId - The ID of the second snapshot to use for the comparison.

The SecondSnapshotId parameter must be specified with a FirstSnapshotID parameter; otherwise, an error occurs.

Request Lenses

listChangedBlocks_nextToken :: Lens' ListChangedBlocks (Maybe Text)

The token to request the next page of results.

listChangedBlocks_startingBlockIndex :: Lens' ListChangedBlocks (Maybe Natural)

The block index from which the comparison should start.

The list in the response will start from this block index or the next valid block index in the snapshots.

listChangedBlocks_firstSnapshotId :: Lens' ListChangedBlocks (Maybe Text)

The ID of the first snapshot to use for the comparison.

The FirstSnapshotID parameter must be specified with a SecondSnapshotId parameter; otherwise, an error occurs.

listChangedBlocks_maxResults :: Lens' ListChangedBlocks (Maybe Natural)

The number of results to return.

listChangedBlocks_secondSnapshotId :: Lens' ListChangedBlocks Text

The ID of the second snapshot to use for the comparison.

The SecondSnapshotId parameter must be specified with a FirstSnapshotID parameter; otherwise, an error occurs.

Destructuring the Response

data ListChangedBlocksResponse

See: newListChangedBlocksResponse smart constructor.

Constructors

ListChangedBlocksResponse'

Fields

blockSize :: Maybe Int

The size of the blocks in the snapshot, in bytes.

expiryTime :: Maybe POSIX

The time when the BlockToken expires.

volumeSize :: Maybe Natural

The size of the volume in GB.

changedBlocks :: Maybe [Sensitive ChangedBlock]

An array of objects containing information about the changed blocks.

nextToken :: Maybe Text

The token to use to retrieve the next page of results. This value is null when there are no more results to return.

httpStatus :: Int

The response's http status code.

newListChangedBlocksResponse

Arguments

:: Int
$sel:httpStatus:ListChangedBlocksResponse'

-> ListChangedBlocksResponse

Create a value of ListChangedBlocksResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:blockSize:ListChangedBlocksResponse', listChangedBlocksResponse_blockSize - The size of the blocks in the snapshot, in bytes.

$sel:expiryTime:ListChangedBlocksResponse', listChangedBlocksResponse_expiryTime - The time when the BlockToken expires.

$sel:volumeSize:ListChangedBlocksResponse', listChangedBlocksResponse_volumeSize - The size of the volume in GB.

$sel:changedBlocks:ListChangedBlocksResponse', listChangedBlocksResponse_changedBlocks - An array of objects containing information about the changed blocks.

$sel:nextToken:ListChangedBlocks', listChangedBlocksResponse_nextToken - The token to use to retrieve the next page of results. This value is null when there are no more results to return.

$sel:httpStatus:ListChangedBlocksResponse', listChangedBlocksResponse_httpStatus - The response's http status code.

Response Lenses

listChangedBlocksResponse_blockSize :: Lens' ListChangedBlocksResponse (Maybe Int)

The size of the blocks in the snapshot, in bytes.

listChangedBlocksResponse_expiryTime :: Lens' ListChangedBlocksResponse (Maybe UTCTime)

The time when the BlockToken expires.

listChangedBlocksResponse_volumeSize :: Lens' ListChangedBlocksResponse (Maybe Natural)

The size of the volume in GB.

listChangedBlocksResponse_changedBlocks :: Lens' ListChangedBlocksResponse (Maybe [ChangedBlock])

An array of objects containing information about the changed blocks.

listChangedBlocksResponse_nextToken :: Lens' ListChangedBlocksResponse (Maybe Text)

The token to use to retrieve the next page of results. This value is null when there are no more results to return.

listChangedBlocksResponse_httpStatus :: Lens' ListChangedBlocksResponse Int

The response's http status code.