Amazonka.EBS.Types

Description

Service Configuration

defaultService :: Service

API version 2019-11-02 of the Amazon Elastic Block Store SDK configuration.

Errors

_ValidationException :: AsError a => Getting (First ServiceError) a ServiceError

The input fails to satisfy the constraints of the EBS direct APIs.

_AccessDeniedException :: AsError a => Getting (First ServiceError) a ServiceError

You do not have sufficient access to perform this action.

_ConflictException :: AsError a => Getting (First ServiceError) a ServiceError

The request uses the same client token as a previous, but non-identical request.

_ServiceQuotaExceededException :: AsError a => Getting (First ServiceError) a ServiceError

Your current service quotas do not allow you to perform this action.

_ConcurrentLimitExceededException :: AsError a => Getting (First ServiceError) a ServiceError

You have reached the limit for concurrent API requests. For more information, see Optimizing performance of the EBS direct APIs in the Amazon Elastic Compute Cloud User Guide.

_RequestThrottledException :: AsError a => Getting (First ServiceError) a ServiceError

The number of API requests has exceed the maximum allowed API request throttling limit.

_InternalServerException :: AsError a => Getting (First ServiceError) a ServiceError

An internal error has occurred.

_ResourceNotFoundException :: AsError a => Getting (First ServiceError) a ServiceError

The specified resource does not exist.

ChecksumAggregationMethod

newtype ChecksumAggregationMethod

Constructors

ChecksumAggregationMethod'

Fields

fromChecksumAggregationMethod :: Text

Bundled Patterns

pattern ChecksumAggregationMethod_LINEAR :: ChecksumAggregationMethod

ChecksumAlgorithm

newtype ChecksumAlgorithm

Constructors

ChecksumAlgorithm'

Fields

fromChecksumAlgorithm :: Text

Bundled Patterns

pattern ChecksumAlgorithm_SHA256 :: ChecksumAlgorithm

Status

newtype Status

Constructors

Status'

Fields

fromStatus :: Text

Bundled Patterns

pattern Status_Completed :: Status

pattern Status_Error :: Status

pattern Status_Pending :: Status

Block

data Block

A block of data in an Amazon Elastic Block Store snapshot.

See: newBlock smart constructor.

Constructors

Block'

Fields

blockIndex :: Maybe Natural

The block index.

blockToken :: Maybe Text

The block token for the block index.

newBlock :: Block

Create a value of Block with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:blockIndex:Block', block_blockIndex - The block index.

$sel:blockToken:Block', block_blockToken - The block token for the block index.

block_blockIndex :: Lens' Block (Maybe Natural)

The block index.

block_blockToken :: Lens' Block (Maybe Text)

The block token for the block index.

ChangedBlock

data ChangedBlock

A block of data in an Amazon Elastic Block Store snapshot that is different from another snapshot of the same volume/snapshot lineage.

See: newChangedBlock smart constructor.

Constructors

ChangedBlock'

Fields

blockIndex :: Maybe Natural

The block index.

secondBlockToken :: Maybe Text

The block token for the block index of the SecondSnapshotId specified in the ListChangedBlocks operation.

firstBlockToken :: Maybe Text

The block token for the block index of the FirstSnapshotId specified in the ListChangedBlocks operation. This value is absent if the first snapshot does not have the changed block that is on the second snapshot.

newChangedBlock :: ChangedBlock

Create a value of ChangedBlock with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:blockIndex:ChangedBlock', changedBlock_blockIndex - The block index.

$sel:secondBlockToken:ChangedBlock', changedBlock_secondBlockToken - The block token for the block index of the SecondSnapshotId specified in the ListChangedBlocks operation.

$sel:firstBlockToken:ChangedBlock', changedBlock_firstBlockToken - The block token for the block index of the FirstSnapshotId specified in the ListChangedBlocks operation. This value is absent if the first snapshot does not have the changed block that is on the second snapshot.

changedBlock_blockIndex :: Lens' ChangedBlock (Maybe Natural)

The block index.

changedBlock_secondBlockToken :: Lens' ChangedBlock (Maybe Text)

The block token for the block index of the SecondSnapshotId specified in the ListChangedBlocks operation.

changedBlock_firstBlockToken :: Lens' ChangedBlock (Maybe Text)

The block token for the block index of the FirstSnapshotId specified in the ListChangedBlocks operation. This value is absent if the first snapshot does not have the changed block that is on the second snapshot.

Tag

data Tag

Describes a tag.

See: newTag smart constructor.

Constructors

Tag'

Fields

value :: Maybe Text

The value of the tag.

key :: Maybe Text

The key of the tag.

newTag :: Tag

Create a value of Tag with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:value:Tag', tag_value - The value of the tag.

$sel:key:Tag', tag_key - The key of the tag.

tag_value :: Lens' Tag (Maybe Text)

The value of the tag.

tag_key :: Lens' Tag (Maybe Text)

The key of the tag.