Amazonka.EC2.AcceptVpcPeeringConnection

Description

Accept a VPC peering connection request. To accept a request, the VPC peering connection must be in the pending-acceptance state, and you must be the owner of the peer VPC. Use DescribeVpcPeeringConnections to view your outstanding VPC peering connection requests.

For an inter-Region VPC peering connection request, you must accept the VPC peering connection in the Region of the accepter VPC.

Creating a Request

data AcceptVpcPeeringConnection

See: newAcceptVpcPeeringConnection smart constructor.

Constructors

AcceptVpcPeeringConnection'

Fields

vpcPeeringConnectionId :: Maybe Text

The ID of the VPC peering connection. You must specify this parameter in the request.

dryRun :: Maybe Bool

Checks whether you have the required permissions for the action, without actually making the request, and provides an error response. If you have the required permissions, the error response is DryRunOperation. Otherwise, it is UnauthorizedOperation.

newAcceptVpcPeeringConnection :: AcceptVpcPeeringConnection

Create a value of AcceptVpcPeeringConnection with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:vpcPeeringConnectionId:AcceptVpcPeeringConnection', acceptVpcPeeringConnection_vpcPeeringConnectionId - The ID of the VPC peering connection. You must specify this parameter in the request.

$sel:dryRun:AcceptVpcPeeringConnection', acceptVpcPeeringConnection_dryRun - Checks whether you have the required permissions for the action, without actually making the request, and provides an error response. If you have the required permissions, the error response is DryRunOperation. Otherwise, it is UnauthorizedOperation.

Request Lenses

acceptVpcPeeringConnection_vpcPeeringConnectionId :: Lens' AcceptVpcPeeringConnection (Maybe Text)

The ID of the VPC peering connection. You must specify this parameter in the request.

acceptVpcPeeringConnection_dryRun :: Lens' AcceptVpcPeeringConnection (Maybe Bool)

Checks whether you have the required permissions for the action, without actually making the request, and provides an error response. If you have the required permissions, the error response is DryRunOperation. Otherwise, it is UnauthorizedOperation.

Destructuring the Response

data AcceptVpcPeeringConnectionResponse

See: newAcceptVpcPeeringConnectionResponse smart constructor.

Constructors

AcceptVpcPeeringConnectionResponse'

Fields

vpcPeeringConnection :: Maybe VpcPeeringConnection

Information about the VPC peering connection.

httpStatus :: Int

The response's http status code.

newAcceptVpcPeeringConnectionResponse

Arguments

:: Int
$sel:httpStatus:AcceptVpcPeeringConnectionResponse'

-> AcceptVpcPeeringConnectionResponse

Create a value of AcceptVpcPeeringConnectionResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:vpcPeeringConnection:AcceptVpcPeeringConnectionResponse', acceptVpcPeeringConnectionResponse_vpcPeeringConnection - Information about the VPC peering connection.

$sel:httpStatus:AcceptVpcPeeringConnectionResponse', acceptVpcPeeringConnectionResponse_httpStatus - The response's http status code.

Response Lenses

acceptVpcPeeringConnectionResponse_vpcPeeringConnection :: Lens' AcceptVpcPeeringConnectionResponse (Maybe VpcPeeringConnection)

Information about the VPC peering connection.

acceptVpcPeeringConnectionResponse_httpStatus :: Lens' AcceptVpcPeeringConnectionResponse Int

The response's http status code.