Amazonka.EC2.AllocateAddress

Description

Allocates an Elastic IP address to your Amazon Web Services account. After you allocate the Elastic IP address you can associate it with an instance or network interface. After you release an Elastic IP address, it is released to the IP address pool and can be allocated to a different Amazon Web Services account.

You can allocate an Elastic IP address from an address pool owned by Amazon Web Services or from an address pool created from a public IPv4 address range that you have brought to Amazon Web Services for use with your Amazon Web Services resources using bring your own IP addresses (BYOIP). For more information, see Bring Your Own IP Addresses (BYOIP) in the Amazon Elastic Compute Cloud User Guide.

EC2-VPC

If you release an Elastic IP address, you might be able to recover it. You cannot recover an Elastic IP address that you released after it is allocated to another Amazon Web Services account. You cannot recover an Elastic IP address for EC2-Classic. To attempt to recover an Elastic IP address that you released, specify it in this operation.

An Elastic IP address is for use either in the EC2-Classic platform or in a VPC. By default, you can allocate 5 Elastic IP addresses for EC2-Classic per Region and 5 Elastic IP addresses for EC2-VPC per Region.

For more information, see Elastic IP Addresses in the Amazon Elastic Compute Cloud User Guide.

You can allocate a carrier IP address which is a public IP address from a telecommunication carrier, to a network interface which resides in a subnet in a Wavelength Zone (for example an EC2 instance).

Creating a Request

data AllocateAddress

See: newAllocateAddress smart constructor.

Constructors

AllocateAddress'

Fields

networkBorderGroup :: Maybe Text

A unique set of Availability Zones, Local Zones, or Wavelength Zones from which Amazon Web Services advertises IP addresses. Use this parameter to limit the IP address to this location. IP addresses cannot move between network border groups.

Use DescribeAvailabilityZones to view the network border groups.

You cannot use a network border group with EC2 Classic. If you attempt this operation on EC2 Classic, you receive an InvalidParameterCombination error.

domain :: Maybe DomainType

Indicates whether the Elastic IP address is for use with instances in a VPC or instances in EC2-Classic.

Default: If the Region supports EC2-Classic, the default is standard. Otherwise, the default is vpc.

address :: Maybe Text

EC2-VPC

The Elastic IP address to recover or an IPv4 address from an address pool.

publicIpv4Pool :: Maybe Text

The ID of an address pool that you own. Use this parameter to let Amazon EC2 select an address from the address pool. To specify a specific address from the address pool, use the Address parameter instead.

tagSpecifications :: Maybe [TagSpecification]

The tags to assign to the Elastic IP address.

customerOwnedIpv4Pool :: Maybe Text

The ID of a customer-owned address pool. Use this parameter to let Amazon EC2 select an address from the address pool. Alternatively, specify a specific address from the address pool.

dryRun :: Maybe Bool

Checks whether you have the required permissions for the action, without actually making the request, and provides an error response. If you have the required permissions, the error response is DryRunOperation. Otherwise, it is UnauthorizedOperation.

newAllocateAddress :: AllocateAddress

Create a value of AllocateAddress with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:networkBorderGroup:AllocateAddress', allocateAddress_networkBorderGroup - A unique set of Availability Zones, Local Zones, or Wavelength Zones from which Amazon Web Services advertises IP addresses. Use this parameter to limit the IP address to this location. IP addresses cannot move between network border groups.

Use DescribeAvailabilityZones to view the network border groups.

You cannot use a network border group with EC2 Classic. If you attempt this operation on EC2 Classic, you receive an InvalidParameterCombination error.

$sel:domain:AllocateAddress', allocateAddress_domain - Indicates whether the Elastic IP address is for use with instances in a VPC or instances in EC2-Classic.

Default: If the Region supports EC2-Classic, the default is standard. Otherwise, the default is vpc.

$sel:address:AllocateAddress', allocateAddress_address - [EC2-VPC] The Elastic IP address to recover or an IPv4 address from an address pool.

$sel:publicIpv4Pool:AllocateAddress', allocateAddress_publicIpv4Pool - The ID of an address pool that you own. Use this parameter to let Amazon EC2 select an address from the address pool. To specify a specific address from the address pool, use the Address parameter instead.

$sel:tagSpecifications:AllocateAddress', allocateAddress_tagSpecifications - The tags to assign to the Elastic IP address.

$sel:customerOwnedIpv4Pool:AllocateAddress', allocateAddress_customerOwnedIpv4Pool - The ID of a customer-owned address pool. Use this parameter to let Amazon EC2 select an address from the address pool. Alternatively, specify a specific address from the address pool.

$sel:dryRun:AllocateAddress', allocateAddress_dryRun - Checks whether you have the required permissions for the action, without actually making the request, and provides an error response. If you have the required permissions, the error response is DryRunOperation. Otherwise, it is UnauthorizedOperation.

Request Lenses

allocateAddress_networkBorderGroup :: Lens' AllocateAddress (Maybe Text)

A unique set of Availability Zones, Local Zones, or Wavelength Zones from which Amazon Web Services advertises IP addresses. Use this parameter to limit the IP address to this location. IP addresses cannot move between network border groups.

Use DescribeAvailabilityZones to view the network border groups.

You cannot use a network border group with EC2 Classic. If you attempt this operation on EC2 Classic, you receive an InvalidParameterCombination error.

allocateAddress_domain :: Lens' AllocateAddress (Maybe DomainType)

Indicates whether the Elastic IP address is for use with instances in a VPC or instances in EC2-Classic.

Default: If the Region supports EC2-Classic, the default is standard. Otherwise, the default is vpc.

allocateAddress_address :: Lens' AllocateAddress (Maybe Text)

EC2-VPC

The Elastic IP address to recover or an IPv4 address from an address pool.

allocateAddress_publicIpv4Pool :: Lens' AllocateAddress (Maybe Text)

The ID of an address pool that you own. Use this parameter to let Amazon EC2 select an address from the address pool. To specify a specific address from the address pool, use the Address parameter instead.

allocateAddress_tagSpecifications :: Lens' AllocateAddress (Maybe [TagSpecification])

The tags to assign to the Elastic IP address.

allocateAddress_customerOwnedIpv4Pool :: Lens' AllocateAddress (Maybe Text)

The ID of a customer-owned address pool. Use this parameter to let Amazon EC2 select an address from the address pool. Alternatively, specify a specific address from the address pool.

allocateAddress_dryRun :: Lens' AllocateAddress (Maybe Bool)

Checks whether you have the required permissions for the action, without actually making the request, and provides an error response. If you have the required permissions, the error response is DryRunOperation. Otherwise, it is UnauthorizedOperation.

Destructuring the Response

data AllocateAddressResponse

See: newAllocateAddressResponse smart constructor.

Constructors

AllocateAddressResponse'

Fields

allocationId :: Maybe Text

EC2-VPC

The ID that Amazon Web Services assigns to represent the allocation of the Elastic IP address for use with instances in a VPC.

carrierIp :: Maybe Text

The carrier IP address. This option is only available for network interfaces which reside in a subnet in a Wavelength Zone (for example an EC2 instance).

networkBorderGroup :: Maybe Text

The set of Availability Zones, Local Zones, or Wavelength Zones from which Amazon Web Services advertises IP addresses.

domain :: Maybe DomainType

Indicates whether the Elastic IP address is for use with instances in a VPC (vpc) or instances in EC2-Classic (standard).

publicIpv4Pool :: Maybe Text

The ID of an address pool.

customerOwnedIpv4Pool :: Maybe Text

The ID of the customer-owned address pool.

customerOwnedIp :: Maybe Text

The customer-owned IP address.

publicIp :: Maybe Text

The Elastic IP address.

httpStatus :: Int

The response's http status code.

newAllocateAddressResponse

Arguments

:: Int
$sel:httpStatus:AllocateAddressResponse'

-> AllocateAddressResponse

Create a value of AllocateAddressResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:allocationId:AllocateAddressResponse', allocateAddressResponse_allocationId - [EC2-VPC] The ID that Amazon Web Services assigns to represent the allocation of the Elastic IP address for use with instances in a VPC.

$sel:carrierIp:AllocateAddressResponse', allocateAddressResponse_carrierIp - The carrier IP address. This option is only available for network interfaces which reside in a subnet in a Wavelength Zone (for example an EC2 instance).

$sel:networkBorderGroup:AllocateAddress', allocateAddressResponse_networkBorderGroup - The set of Availability Zones, Local Zones, or Wavelength Zones from which Amazon Web Services advertises IP addresses.

$sel:domain:AllocateAddress', allocateAddressResponse_domain - Indicates whether the Elastic IP address is for use with instances in a VPC (vpc) or instances in EC2-Classic (standard).

$sel:publicIpv4Pool:AllocateAddress', allocateAddressResponse_publicIpv4Pool - The ID of an address pool.

$sel:customerOwnedIpv4Pool:AllocateAddress', allocateAddressResponse_customerOwnedIpv4Pool - The ID of the customer-owned address pool.

$sel:customerOwnedIp:AllocateAddressResponse', allocateAddressResponse_customerOwnedIp - The customer-owned IP address.

$sel:publicIp:AllocateAddressResponse', allocateAddressResponse_publicIp - The Elastic IP address.

$sel:httpStatus:AllocateAddressResponse', allocateAddressResponse_httpStatus - The response's http status code.

Response Lenses

allocateAddressResponse_allocationId :: Lens' AllocateAddressResponse (Maybe Text)

EC2-VPC

The ID that Amazon Web Services assigns to represent the allocation of the Elastic IP address for use with instances in a VPC.

allocateAddressResponse_carrierIp :: Lens' AllocateAddressResponse (Maybe Text)

The carrier IP address. This option is only available for network interfaces which reside in a subnet in a Wavelength Zone (for example an EC2 instance).

allocateAddressResponse_networkBorderGroup :: Lens' AllocateAddressResponse (Maybe Text)

The set of Availability Zones, Local Zones, or Wavelength Zones from which Amazon Web Services advertises IP addresses.

allocateAddressResponse_domain :: Lens' AllocateAddressResponse (Maybe DomainType)

Indicates whether the Elastic IP address is for use with instances in a VPC (vpc) or instances in EC2-Classic (standard).

allocateAddressResponse_publicIpv4Pool :: Lens' AllocateAddressResponse (Maybe Text)

The ID of an address pool.

allocateAddressResponse_customerOwnedIpv4Pool :: Lens' AllocateAddressResponse (Maybe Text)

The ID of the customer-owned address pool.

allocateAddressResponse_customerOwnedIp :: Lens' AllocateAddressResponse (Maybe Text)

The customer-owned IP address.

allocateAddressResponse_publicIp :: Lens' AllocateAddressResponse (Maybe Text)

The Elastic IP address.

allocateAddressResponse_httpStatus :: Lens' AllocateAddressResponse Int

The response's http status code.