Amazonka.EC2.AllocateHosts

Description

Allocates a Dedicated Host to your account. At a minimum, specify the supported instance type or instance family, the Availability Zone in which to allocate the host, and the number of hosts to allocate.

Creating a Request

data AllocateHosts

See: newAllocateHosts smart constructor.

Constructors

AllocateHosts'

Fields

instanceFamily :: Maybe Text

Specifies the instance family to be supported by the Dedicated Hosts. If you specify an instance family, the Dedicated Hosts support multiple instance types within that instance family.

If you want the Dedicated Hosts to support a specific instance type only, omit this parameter and specify InstanceType instead. You cannot specify InstanceFamily and InstanceType in the same request.

clientToken :: Maybe Text

Unique, case-sensitive identifier that you provide to ensure the idempotency of the request. For more information, see Ensuring Idempotency.

instanceType :: Maybe Text

Specifies the instance type to be supported by the Dedicated Hosts. If you specify an instance type, the Dedicated Hosts support instances of the specified instance type only.

If you want the Dedicated Hosts to support multiple instance types in a specific instance family, omit this parameter and specify InstanceFamily instead. You cannot specify InstanceType and InstanceFamily in the same request.

tagSpecifications :: Maybe [TagSpecification]

The tags to apply to the Dedicated Host during creation.

hostRecovery :: Maybe HostRecovery

Indicates whether to enable or disable host recovery for the Dedicated Host. Host recovery is disabled by default. For more information, see Host recovery in the Amazon EC2 User Guide.

Default: off

autoPlacement :: Maybe AutoPlacement

Indicates whether the host accepts any untargeted instance launches that match its instance type configuration, or if it only accepts Host tenancy instance launches that specify its unique host ID. For more information, see Understanding auto-placement and affinity in the Amazon EC2 User Guide.

Default: on

availabilityZone :: Text

The Availability Zone in which to allocate the Dedicated Host.

quantity :: Int

The number of Dedicated Hosts to allocate to your account with these parameters.

newAllocateHosts

Arguments

:: Text
$sel:availabilityZone:AllocateHosts'

-> Int
$sel:quantity:AllocateHosts'

-> AllocateHosts

Create a value of AllocateHosts with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:instanceFamily:AllocateHosts', allocateHosts_instanceFamily - Specifies the instance family to be supported by the Dedicated Hosts. If you specify an instance family, the Dedicated Hosts support multiple instance types within that instance family.

If you want the Dedicated Hosts to support a specific instance type only, omit this parameter and specify InstanceType instead. You cannot specify InstanceFamily and InstanceType in the same request.

$sel:clientToken:AllocateHosts', allocateHosts_clientToken - Unique, case-sensitive identifier that you provide to ensure the idempotency of the request. For more information, see Ensuring Idempotency.

$sel:instanceType:AllocateHosts', allocateHosts_instanceType - Specifies the instance type to be supported by the Dedicated Hosts. If you specify an instance type, the Dedicated Hosts support instances of the specified instance type only.

If you want the Dedicated Hosts to support multiple instance types in a specific instance family, omit this parameter and specify InstanceFamily instead. You cannot specify InstanceType and InstanceFamily in the same request.

$sel:tagSpecifications:AllocateHosts', allocateHosts_tagSpecifications - The tags to apply to the Dedicated Host during creation.

$sel:hostRecovery:AllocateHosts', allocateHosts_hostRecovery - Indicates whether to enable or disable host recovery for the Dedicated Host. Host recovery is disabled by default. For more information, see Host recovery in the Amazon EC2 User Guide.

Default: off

$sel:autoPlacement:AllocateHosts', allocateHosts_autoPlacement - Indicates whether the host accepts any untargeted instance launches that match its instance type configuration, or if it only accepts Host tenancy instance launches that specify its unique host ID. For more information, see Understanding auto-placement and affinity in the Amazon EC2 User Guide.

Default: on

$sel:availabilityZone:AllocateHosts', allocateHosts_availabilityZone - The Availability Zone in which to allocate the Dedicated Host.

$sel:quantity:AllocateHosts', allocateHosts_quantity - The number of Dedicated Hosts to allocate to your account with these parameters.

Request Lenses

allocateHosts_instanceFamily :: Lens' AllocateHosts (Maybe Text)

Specifies the instance family to be supported by the Dedicated Hosts. If you specify an instance family, the Dedicated Hosts support multiple instance types within that instance family.

If you want the Dedicated Hosts to support a specific instance type only, omit this parameter and specify InstanceType instead. You cannot specify InstanceFamily and InstanceType in the same request.

allocateHosts_clientToken :: Lens' AllocateHosts (Maybe Text)

Unique, case-sensitive identifier that you provide to ensure the idempotency of the request. For more information, see Ensuring Idempotency.

allocateHosts_instanceType :: Lens' AllocateHosts (Maybe Text)

Specifies the instance type to be supported by the Dedicated Hosts. If you specify an instance type, the Dedicated Hosts support instances of the specified instance type only.

If you want the Dedicated Hosts to support multiple instance types in a specific instance family, omit this parameter and specify InstanceFamily instead. You cannot specify InstanceType and InstanceFamily in the same request.

allocateHosts_tagSpecifications :: Lens' AllocateHosts (Maybe [TagSpecification])

The tags to apply to the Dedicated Host during creation.

allocateHosts_hostRecovery :: Lens' AllocateHosts (Maybe HostRecovery)

Indicates whether to enable or disable host recovery for the Dedicated Host. Host recovery is disabled by default. For more information, see Host recovery in the Amazon EC2 User Guide.

Default: off

allocateHosts_autoPlacement :: Lens' AllocateHosts (Maybe AutoPlacement)

Indicates whether the host accepts any untargeted instance launches that match its instance type configuration, or if it only accepts Host tenancy instance launches that specify its unique host ID. For more information, see Understanding auto-placement and affinity in the Amazon EC2 User Guide.

Default: on

allocateHosts_availabilityZone :: Lens' AllocateHosts Text

The Availability Zone in which to allocate the Dedicated Host.

allocateHosts_quantity :: Lens' AllocateHosts Int

The number of Dedicated Hosts to allocate to your account with these parameters.

Destructuring the Response

data AllocateHostsResponse

Contains the output of AllocateHosts.

See: newAllocateHostsResponse smart constructor.

Constructors

AllocateHostsResponse'

Fields

hostIds :: Maybe [Text]

The ID of the allocated Dedicated Host. This is used to launch an instance onto a specific host.

httpStatus :: Int

The response's http status code.

newAllocateHostsResponse

Arguments

:: Int
$sel:httpStatus:AllocateHostsResponse'

-> AllocateHostsResponse

Create a value of AllocateHostsResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:hostIds:AllocateHostsResponse', allocateHostsResponse_hostIds - The ID of the allocated Dedicated Host. This is used to launch an instance onto a specific host.

$sel:httpStatus:AllocateHostsResponse', allocateHostsResponse_httpStatus - The response's http status code.

Response Lenses

allocateHostsResponse_hostIds :: Lens' AllocateHostsResponse (Maybe [Text])

The ID of the allocated Dedicated Host. This is used to launch an instance onto a specific host.

allocateHostsResponse_httpStatus :: Lens' AllocateHostsResponse Int

The response's http status code.