Amazonka.EC2.AssociateAddress

Description

Associates an Elastic IP address, or carrier IP address (for instances that are in subnets in Wavelength Zones) with an instance or a network interface. Before you can use an Elastic IP address, you must allocate it to your account.

An Elastic IP address is for use in either the EC2-Classic platform or in a VPC. For more information, see Elastic IP Addresses in the Amazon Elastic Compute Cloud User Guide.

EC2-Classic, VPC in an EC2-VPC-only account

If the Elastic IP address is already associated with a different instance, it is disassociated from that instance and associated with the specified instance. If you associate an Elastic IP address with an instance that has an existing Elastic IP address, the existing address is disassociated from the instance, but remains allocated to your account.

VPC in an EC2-Classic account

If you don't specify a private IP address, the Elastic IP address is associated with the primary IP address. If the Elastic IP address is already associated with a different instance or a network interface, you get an error unless you allow reassociation. You cannot associate an Elastic IP address with an instance or network interface that has an existing Elastic IP address.

Subnets in Wavelength Zones

You can associate an IP address from the telecommunication carrier to the instance or network interface.

You cannot associate an Elastic IP address with an interface in a different network border group.

This is an idempotent operation. If you perform the operation more than once, Amazon EC2 doesn't return an error, and you may be charged for each time the Elastic IP address is remapped to the same instance. For more information, see the Elastic IP Addresses section of Amazon EC2 Pricing.

Creating a Request

data AssociateAddress

See: newAssociateAddress smart constructor.

Constructors

AssociateAddress'

Fields

instanceId :: Maybe Text

The ID of the instance. The instance must have exactly one attached network interface. For EC2-VPC, you can specify either the instance ID or the network interface ID, but not both. For EC2-Classic, you must specify an instance ID and the instance must be in the running state.

allocationId :: Maybe Text

EC2-VPC

The allocation ID. This is required for EC2-VPC.

networkInterfaceId :: Maybe Text

EC2-VPC

The ID of the network interface. If the instance has more than one network interface, you must specify a network interface ID.

For EC2-VPC, you can specify either the instance ID or the network interface ID, but not both.

allowReassociation :: Maybe Bool

EC2-VPC

For a VPC in an EC2-Classic account, specify true to allow an Elastic IP address that is already associated with an instance or network interface to be reassociated with the specified instance or network interface. Otherwise, the operation fails. In a VPC in an EC2-VPC-only account, reassociation is automatic, therefore you can specify false to ensure the operation fails if the Elastic IP address is already associated with another resource.

privateIpAddress :: Maybe Text

EC2-VPC

The primary or secondary private IP address to associate with the Elastic IP address. If no private IP address is specified, the Elastic IP address is associated with the primary private IP address.

publicIp :: Maybe Text

EC2-Classic

The Elastic IP address to associate with the instance. This is required for EC2-Classic.

dryRun :: Maybe Bool

Checks whether you have the required permissions for the action, without actually making the request, and provides an error response. If you have the required permissions, the error response is DryRunOperation. Otherwise, it is UnauthorizedOperation.

newAssociateAddress :: AssociateAddress

Create a value of AssociateAddress with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:instanceId:AssociateAddress', associateAddress_instanceId - The ID of the instance. The instance must have exactly one attached network interface. For EC2-VPC, you can specify either the instance ID or the network interface ID, but not both. For EC2-Classic, you must specify an instance ID and the instance must be in the running state.

$sel:allocationId:AssociateAddress', associateAddress_allocationId - [EC2-VPC] The allocation ID. This is required for EC2-VPC.

$sel:networkInterfaceId:AssociateAddress', associateAddress_networkInterfaceId - [EC2-VPC] The ID of the network interface. If the instance has more than one network interface, you must specify a network interface ID.

For EC2-VPC, you can specify either the instance ID or the network interface ID, but not both.

$sel:allowReassociation:AssociateAddress', associateAddress_allowReassociation - [EC2-VPC] For a VPC in an EC2-Classic account, specify true to allow an Elastic IP address that is already associated with an instance or network interface to be reassociated with the specified instance or network interface. Otherwise, the operation fails. In a VPC in an EC2-VPC-only account, reassociation is automatic, therefore you can specify false to ensure the operation fails if the Elastic IP address is already associated with another resource.

$sel:privateIpAddress:AssociateAddress', associateAddress_privateIpAddress - [EC2-VPC] The primary or secondary private IP address to associate with the Elastic IP address. If no private IP address is specified, the Elastic IP address is associated with the primary private IP address.

$sel:publicIp:AssociateAddress', associateAddress_publicIp - [EC2-Classic] The Elastic IP address to associate with the instance. This is required for EC2-Classic.

$sel:dryRun:AssociateAddress', associateAddress_dryRun - Checks whether you have the required permissions for the action, without actually making the request, and provides an error response. If you have the required permissions, the error response is DryRunOperation. Otherwise, it is UnauthorizedOperation.

Request Lenses

associateAddress_instanceId :: Lens' AssociateAddress (Maybe Text)

The ID of the instance. The instance must have exactly one attached network interface. For EC2-VPC, you can specify either the instance ID or the network interface ID, but not both. For EC2-Classic, you must specify an instance ID and the instance must be in the running state.

associateAddress_allocationId :: Lens' AssociateAddress (Maybe Text)

EC2-VPC

The allocation ID. This is required for EC2-VPC.

associateAddress_networkInterfaceId :: Lens' AssociateAddress (Maybe Text)

EC2-VPC

The ID of the network interface. If the instance has more than one network interface, you must specify a network interface ID.

For EC2-VPC, you can specify either the instance ID or the network interface ID, but not both.

associateAddress_allowReassociation :: Lens' AssociateAddress (Maybe Bool)

EC2-VPC

For a VPC in an EC2-Classic account, specify true to allow an Elastic IP address that is already associated with an instance or network interface to be reassociated with the specified instance or network interface. Otherwise, the operation fails. In a VPC in an EC2-VPC-only account, reassociation is automatic, therefore you can specify false to ensure the operation fails if the Elastic IP address is already associated with another resource.

associateAddress_privateIpAddress :: Lens' AssociateAddress (Maybe Text)

EC2-VPC

The primary or secondary private IP address to associate with the Elastic IP address. If no private IP address is specified, the Elastic IP address is associated with the primary private IP address.

associateAddress_publicIp :: Lens' AssociateAddress (Maybe Text)

EC2-Classic

The Elastic IP address to associate with the instance. This is required for EC2-Classic.

associateAddress_dryRun :: Lens' AssociateAddress (Maybe Bool)

Checks whether you have the required permissions for the action, without actually making the request, and provides an error response. If you have the required permissions, the error response is DryRunOperation. Otherwise, it is UnauthorizedOperation.

Destructuring the Response

data AssociateAddressResponse

See: newAssociateAddressResponse smart constructor.

Constructors

AssociateAddressResponse'

Fields

associationId :: Maybe Text

EC2-VPC

The ID that represents the association of the Elastic IP address with an instance.

httpStatus :: Int

The response's http status code.

newAssociateAddressResponse

Arguments

:: Int
$sel:httpStatus:AssociateAddressResponse'

-> AssociateAddressResponse

Create a value of AssociateAddressResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:associationId:AssociateAddressResponse', associateAddressResponse_associationId - [EC2-VPC] The ID that represents the association of the Elastic IP address with an instance.

$sel:httpStatus:AssociateAddressResponse', associateAddressResponse_httpStatus - The response's http status code.

Response Lenses

associateAddressResponse_associationId :: Lens' AssociateAddressResponse (Maybe Text)

EC2-VPC

The ID that represents the association of the Elastic IP address with an instance.

associateAddressResponse_httpStatus :: Lens' AssociateAddressResponse Int

The response's http status code.