Amazonka.EC2.AttachInternetGateway

Description

Attaches an internet gateway or a virtual private gateway to a VPC, enabling connectivity between the internet and the VPC. For more information about your VPC and internet gateway, see the Amazon Virtual Private Cloud User Guide.

Creating a Request

data AttachInternetGateway

See: newAttachInternetGateway smart constructor.

Constructors

AttachInternetGateway'

Fields

dryRun :: Maybe Bool

Checks whether you have the required permissions for the action, without actually making the request, and provides an error response. If you have the required permissions, the error response is DryRunOperation. Otherwise, it is UnauthorizedOperation.

internetGatewayId :: Text

The ID of the internet gateway.

vpcId :: Text

The ID of the VPC.

newAttachInternetGateway

Arguments

:: Text
$sel:internetGatewayId:AttachInternetGateway'

-> Text
$sel:vpcId:AttachInternetGateway'

-> AttachInternetGateway

Create a value of AttachInternetGateway with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:dryRun:AttachInternetGateway', attachInternetGateway_dryRun - Checks whether you have the required permissions for the action, without actually making the request, and provides an error response. If you have the required permissions, the error response is DryRunOperation. Otherwise, it is UnauthorizedOperation.

$sel:internetGatewayId:AttachInternetGateway', attachInternetGateway_internetGatewayId - The ID of the internet gateway.

$sel:vpcId:AttachInternetGateway', attachInternetGateway_vpcId - The ID of the VPC.

Request Lenses

attachInternetGateway_dryRun :: Lens' AttachInternetGateway (Maybe Bool)

Checks whether you have the required permissions for the action, without actually making the request, and provides an error response. If you have the required permissions, the error response is DryRunOperation. Otherwise, it is UnauthorizedOperation.

attachInternetGateway_internetGatewayId :: Lens' AttachInternetGateway Text

The ID of the internet gateway.

attachInternetGateway_vpcId :: Lens' AttachInternetGateway Text

The ID of the VPC.

Destructuring the Response

data AttachInternetGatewayResponse

See: newAttachInternetGatewayResponse smart constructor.

Constructors

AttachInternetGatewayResponse'

newAttachInternetGatewayResponse :: AttachInternetGatewayResponse

Create a value of AttachInternetGatewayResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.