Amazonka.EC2.CreateDefaultSubnet

Description

Creates a default subnet with a size /20 IPv4 CIDR block in the specified Availability Zone in your default VPC. You can have only one default subnet per Availability Zone. For more information, see Creating a default subnet in the Amazon Virtual Private Cloud User Guide.

Creating a Request

data CreateDefaultSubnet

See: newCreateDefaultSubnet smart constructor.

Constructors

CreateDefaultSubnet'

Fields

dryRun :: Maybe Bool

Checks whether you have the required permissions for the action, without actually making the request, and provides an error response. If you have the required permissions, the error response is DryRunOperation. Otherwise, it is UnauthorizedOperation.

availabilityZone :: Text

The Availability Zone in which to create the default subnet.

newCreateDefaultSubnet

Arguments

:: Text
$sel:availabilityZone:CreateDefaultSubnet'

-> CreateDefaultSubnet

Create a value of CreateDefaultSubnet with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:dryRun:CreateDefaultSubnet', createDefaultSubnet_dryRun - Checks whether you have the required permissions for the action, without actually making the request, and provides an error response. If you have the required permissions, the error response is DryRunOperation. Otherwise, it is UnauthorizedOperation.

$sel:availabilityZone:CreateDefaultSubnet', createDefaultSubnet_availabilityZone - The Availability Zone in which to create the default subnet.

Request Lenses

createDefaultSubnet_dryRun :: Lens' CreateDefaultSubnet (Maybe Bool)

Checks whether you have the required permissions for the action, without actually making the request, and provides an error response. If you have the required permissions, the error response is DryRunOperation. Otherwise, it is UnauthorizedOperation.

createDefaultSubnet_availabilityZone :: Lens' CreateDefaultSubnet Text

The Availability Zone in which to create the default subnet.

Destructuring the Response

data CreateDefaultSubnetResponse

See: newCreateDefaultSubnetResponse smart constructor.

Constructors

CreateDefaultSubnetResponse'

Fields

subnet :: Maybe Subnet

Information about the subnet.

httpStatus :: Int

The response's http status code.

newCreateDefaultSubnetResponse

Arguments

:: Int
$sel:httpStatus:CreateDefaultSubnetResponse'

-> CreateDefaultSubnetResponse

Create a value of CreateDefaultSubnetResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:subnet:CreateDefaultSubnetResponse', createDefaultSubnetResponse_subnet - Information about the subnet.

$sel:httpStatus:CreateDefaultSubnetResponse', createDefaultSubnetResponse_httpStatus - The response's http status code.

Response Lenses

createDefaultSubnetResponse_subnet :: Lens' CreateDefaultSubnetResponse (Maybe Subnet)

Information about the subnet.

createDefaultSubnetResponse_httpStatus :: Lens' CreateDefaultSubnetResponse Int

The response's http status code.