Amazonka.EC2.CreateEgressOnlyInternetGateway

Description

IPv6 only

Creates an egress-only internet gateway for your VPC. An egress-only internet gateway is used to enable outbound communication over IPv6 from instances in your VPC to the internet, and prevents hosts outside of your VPC from initiating an IPv6 connection with your instance.

Creating a Request

data CreateEgressOnlyInternetGateway

See: newCreateEgressOnlyInternetGateway smart constructor.

Constructors

CreateEgressOnlyInternetGateway'

Fields

clientToken :: Maybe Text

Unique, case-sensitive identifier that you provide to ensure the idempotency of the request. For more information, see How to ensure idempotency.

tagSpecifications :: Maybe [TagSpecification]

The tags to assign to the egress-only internet gateway.

dryRun :: Maybe Bool

Checks whether you have the required permissions for the action, without actually making the request, and provides an error response. If you have the required permissions, the error response is DryRunOperation. Otherwise, it is UnauthorizedOperation.

vpcId :: Text

The ID of the VPC for which to create the egress-only internet gateway.

newCreateEgressOnlyInternetGateway

Arguments

:: Text
$sel:vpcId:CreateEgressOnlyInternetGateway'

-> CreateEgressOnlyInternetGateway

Create a value of CreateEgressOnlyInternetGateway with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:clientToken:CreateEgressOnlyInternetGateway', createEgressOnlyInternetGateway_clientToken - Unique, case-sensitive identifier that you provide to ensure the idempotency of the request. For more information, see How to ensure idempotency.

$sel:tagSpecifications:CreateEgressOnlyInternetGateway', createEgressOnlyInternetGateway_tagSpecifications - The tags to assign to the egress-only internet gateway.

$sel:dryRun:CreateEgressOnlyInternetGateway', createEgressOnlyInternetGateway_dryRun - Checks whether you have the required permissions for the action, without actually making the request, and provides an error response. If you have the required permissions, the error response is DryRunOperation. Otherwise, it is UnauthorizedOperation.

$sel:vpcId:CreateEgressOnlyInternetGateway', createEgressOnlyInternetGateway_vpcId - The ID of the VPC for which to create the egress-only internet gateway.

Request Lenses

createEgressOnlyInternetGateway_clientToken :: Lens' CreateEgressOnlyInternetGateway (Maybe Text)

Unique, case-sensitive identifier that you provide to ensure the idempotency of the request. For more information, see How to ensure idempotency.

createEgressOnlyInternetGateway_tagSpecifications :: Lens' CreateEgressOnlyInternetGateway (Maybe [TagSpecification])

The tags to assign to the egress-only internet gateway.

createEgressOnlyInternetGateway_dryRun :: Lens' CreateEgressOnlyInternetGateway (Maybe Bool)

Checks whether you have the required permissions for the action, without actually making the request, and provides an error response. If you have the required permissions, the error response is DryRunOperation. Otherwise, it is UnauthorizedOperation.

createEgressOnlyInternetGateway_vpcId :: Lens' CreateEgressOnlyInternetGateway Text

The ID of the VPC for which to create the egress-only internet gateway.

Destructuring the Response

data CreateEgressOnlyInternetGatewayResponse

See: newCreateEgressOnlyInternetGatewayResponse smart constructor.

Constructors

CreateEgressOnlyInternetGatewayResponse'

Fields

clientToken :: Maybe Text

Unique, case-sensitive identifier that you provide to ensure the idempotency of the request.

egressOnlyInternetGateway :: Maybe EgressOnlyInternetGateway

Information about the egress-only internet gateway.

httpStatus :: Int

The response's http status code.

newCreateEgressOnlyInternetGatewayResponse

Arguments

:: Int
$sel:httpStatus:CreateEgressOnlyInternetGatewayResponse'

-> CreateEgressOnlyInternetGatewayResponse

Create a value of CreateEgressOnlyInternetGatewayResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:clientToken:CreateEgressOnlyInternetGateway', createEgressOnlyInternetGatewayResponse_clientToken - Unique, case-sensitive identifier that you provide to ensure the idempotency of the request.

$sel:egressOnlyInternetGateway:CreateEgressOnlyInternetGatewayResponse', createEgressOnlyInternetGatewayResponse_egressOnlyInternetGateway - Information about the egress-only internet gateway.

$sel:httpStatus:CreateEgressOnlyInternetGatewayResponse', createEgressOnlyInternetGatewayResponse_httpStatus - The response's http status code.

Response Lenses

createEgressOnlyInternetGatewayResponse_clientToken :: Lens' CreateEgressOnlyInternetGatewayResponse (Maybe Text)

Unique, case-sensitive identifier that you provide to ensure the idempotency of the request.

createEgressOnlyInternetGatewayResponse_egressOnlyInternetGateway :: Lens' CreateEgressOnlyInternetGatewayResponse (Maybe EgressOnlyInternetGateway)

Information about the egress-only internet gateway.

createEgressOnlyInternetGatewayResponse_httpStatus :: Lens' CreateEgressOnlyInternetGatewayResponse Int

The response's http status code.