Amazonka.EC2.CreateImage

Description

Creates an Amazon EBS-backed AMI from an Amazon EBS-backed instance that is either running or stopped.

If you customized your instance with instance store volumes or Amazon EBS volumes in addition to the root device volume, the new AMI contains block device mapping information for those volumes. When you launch an instance from this new AMI, the instance automatically launches with those additional volumes.

For more information, see Creating Amazon EBS-Backed Linux AMIs in the Amazon Elastic Compute Cloud User Guide.

Creating a Request

data CreateImage

See: newCreateImage smart constructor.

Constructors

CreateImage'

Fields

tagSpecifications :: Maybe [TagSpecification]

The tags to apply to the AMI and snapshots on creation. You can tag the AMI, the snapshots, or both.

To tag the AMI, the value for ResourceType must be image.

To tag the snapshots that are created of the root volume and of other Amazon EBS volumes that are attached to the instance, the value for ResourceType must be snapshot. The same tag is applied to all of the snapshots that are created.

If you specify other values for ResourceType, the request fails.

To tag an AMI or snapshot after it has been created, see CreateTags.

noReboot :: Maybe Bool

By default, Amazon EC2 attempts to shut down and reboot the instance before creating the image. If the No Reboot option is set, Amazon EC2 doesn't shut down the instance before creating the image. Without a reboot, the AMI will be crash consistent (all the volumes are snapshotted at the same time), but not application consistent (all the operating system buffers are not flushed to disk before the snapshots are created).

description :: Maybe Text

A description for the new image.

blockDeviceMappings :: Maybe [BlockDeviceMapping]

The block device mappings. This parameter cannot be used to modify the encryption status of existing volumes or snapshots. To create an AMI with encrypted snapshots, use the CopyImage action.

dryRun :: Maybe Bool

Checks whether you have the required permissions for the action, without actually making the request, and provides an error response. If you have the required permissions, the error response is DryRunOperation. Otherwise, it is UnauthorizedOperation.

instanceId :: Text

The ID of the instance.

name :: Text

A name for the new image.

Constraints: 3-128 alphanumeric characters, parentheses (()), square brackets ([]), spaces ( ), periods (.), slashes (/), dashes (-), single quotes ('), at-signs (@), or underscores(_)

newCreateImage

Arguments

:: Text
$sel:instanceId:CreateImage'

-> Text
$sel:name:CreateImage'

-> CreateImage

Create a value of CreateImage with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:tagSpecifications:CreateImage', createImage_tagSpecifications - The tags to apply to the AMI and snapshots on creation. You can tag the AMI, the snapshots, or both.

To tag the AMI, the value for ResourceType must be image.

To tag the snapshots that are created of the root volume and of other Amazon EBS volumes that are attached to the instance, the value for ResourceType must be snapshot. The same tag is applied to all of the snapshots that are created.

If you specify other values for ResourceType, the request fails.

To tag an AMI or snapshot after it has been created, see CreateTags.

$sel:noReboot:CreateImage', createImage_noReboot - By default, Amazon EC2 attempts to shut down and reboot the instance before creating the image. If the No Reboot option is set, Amazon EC2 doesn't shut down the instance before creating the image. Without a reboot, the AMI will be crash consistent (all the volumes are snapshotted at the same time), but not application consistent (all the operating system buffers are not flushed to disk before the snapshots are created).

$sel:description:CreateImage', createImage_description - A description for the new image.

$sel:blockDeviceMappings:CreateImage', createImage_blockDeviceMappings - The block device mappings. This parameter cannot be used to modify the encryption status of existing volumes or snapshots. To create an AMI with encrypted snapshots, use the CopyImage action.

$sel:dryRun:CreateImage', createImage_dryRun - Checks whether you have the required permissions for the action, without actually making the request, and provides an error response. If you have the required permissions, the error response is DryRunOperation. Otherwise, it is UnauthorizedOperation.

$sel:instanceId:CreateImage', createImage_instanceId - The ID of the instance.

$sel:name:CreateImage', createImage_name - A name for the new image.

Constraints: 3-128 alphanumeric characters, parentheses (()), square brackets ([]), spaces ( ), periods (.), slashes (/), dashes (-), single quotes ('), at-signs (@), or underscores(_)

Request Lenses

createImage_tagSpecifications :: Lens' CreateImage (Maybe [TagSpecification])

The tags to apply to the AMI and snapshots on creation. You can tag the AMI, the snapshots, or both.

To tag the AMI, the value for ResourceType must be image.

To tag the snapshots that are created of the root volume and of other Amazon EBS volumes that are attached to the instance, the value for ResourceType must be snapshot. The same tag is applied to all of the snapshots that are created.

If you specify other values for ResourceType, the request fails.

To tag an AMI or snapshot after it has been created, see CreateTags.

createImage_noReboot :: Lens' CreateImage (Maybe Bool)

By default, Amazon EC2 attempts to shut down and reboot the instance before creating the image. If the No Reboot option is set, Amazon EC2 doesn't shut down the instance before creating the image. Without a reboot, the AMI will be crash consistent (all the volumes are snapshotted at the same time), but not application consistent (all the operating system buffers are not flushed to disk before the snapshots are created).

createImage_description :: Lens' CreateImage (Maybe Text)

A description for the new image.

createImage_blockDeviceMappings :: Lens' CreateImage (Maybe [BlockDeviceMapping])

The block device mappings. This parameter cannot be used to modify the encryption status of existing volumes or snapshots. To create an AMI with encrypted snapshots, use the CopyImage action.

createImage_dryRun :: Lens' CreateImage (Maybe Bool)

Checks whether you have the required permissions for the action, without actually making the request, and provides an error response. If you have the required permissions, the error response is DryRunOperation. Otherwise, it is UnauthorizedOperation.

createImage_instanceId :: Lens' CreateImage Text

The ID of the instance.

createImage_name :: Lens' CreateImage Text

A name for the new image.

Constraints: 3-128 alphanumeric characters, parentheses (()), square brackets ([]), spaces ( ), periods (.), slashes (/), dashes (-), single quotes ('), at-signs (@), or underscores(_)

Destructuring the Response

data CreateImageResponse

See: newCreateImageResponse smart constructor.

Constructors

CreateImageResponse'

Fields

imageId :: Maybe Text

The ID of the new AMI.

httpStatus :: Int

The response's http status code.

newCreateImageResponse

Arguments

:: Int
$sel:httpStatus:CreateImageResponse'

-> CreateImageResponse

Create a value of CreateImageResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:imageId:CreateImageResponse', createImageResponse_imageId - The ID of the new AMI.

$sel:httpStatus:CreateImageResponse', createImageResponse_httpStatus - The response's http status code.

Response Lenses

createImageResponse_imageId :: Lens' CreateImageResponse (Maybe Text)

The ID of the new AMI.

createImageResponse_httpStatus :: Lens' CreateImageResponse Int

The response's http status code.