Amazonka.EC2.CreateKeyPair

Description

Creates an ED25519 or 2048-bit RSA key pair with the specified name. Amazon EC2 stores the public key and displays the private key for you to save to a file. The private key is returned as an unencrypted PEM encoded PKCS#1 private key. If a key with the specified name already exists, Amazon EC2 returns an error.

The key pair returned to you is available only in the Amazon Web Services Region in which you create it. If you prefer, you can create your own key pair using a third-party tool and upload it to any Region using ImportKeyPair.

You can have up to 5,000 key pairs per Amazon Web Services Region.

For more information, see Amazon EC2 key pairs in the Amazon Elastic Compute Cloud User Guide.

Creating a Request

data CreateKeyPair

See: newCreateKeyPair smart constructor.

Constructors

CreateKeyPair'

Fields

keyType :: Maybe KeyType

The type of key pair. Note that ED25519 keys are not supported for Windows instances, EC2 Instance Connect, and EC2 Serial Console.

Default: rsa

tagSpecifications :: Maybe [TagSpecification]

The tags to apply to the new key pair.

dryRun :: Maybe Bool

Checks whether you have the required permissions for the action, without actually making the request, and provides an error response. If you have the required permissions, the error response is DryRunOperation. Otherwise, it is UnauthorizedOperation.

keyName :: Text

A unique name for the key pair.

Constraints: Up to 255 ASCII characters

newCreateKeyPair

Arguments

:: Text
$sel:keyName:CreateKeyPair'

-> CreateKeyPair

Create a value of CreateKeyPair with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:keyType:CreateKeyPair', createKeyPair_keyType - The type of key pair. Note that ED25519 keys are not supported for Windows instances, EC2 Instance Connect, and EC2 Serial Console.

Default: rsa

$sel:tagSpecifications:CreateKeyPair', createKeyPair_tagSpecifications - The tags to apply to the new key pair.

$sel:dryRun:CreateKeyPair', createKeyPair_dryRun - Checks whether you have the required permissions for the action, without actually making the request, and provides an error response. If you have the required permissions, the error response is DryRunOperation. Otherwise, it is UnauthorizedOperation.

$sel:keyName:CreateKeyPair', createKeyPair_keyName - A unique name for the key pair.

Constraints: Up to 255 ASCII characters

Request Lenses

createKeyPair_keyType :: Lens' CreateKeyPair (Maybe KeyType)

The type of key pair. Note that ED25519 keys are not supported for Windows instances, EC2 Instance Connect, and EC2 Serial Console.

Default: rsa

createKeyPair_tagSpecifications :: Lens' CreateKeyPair (Maybe [TagSpecification])

The tags to apply to the new key pair.

createKeyPair_dryRun :: Lens' CreateKeyPair (Maybe Bool)

Checks whether you have the required permissions for the action, without actually making the request, and provides an error response. If you have the required permissions, the error response is DryRunOperation. Otherwise, it is UnauthorizedOperation.

createKeyPair_keyName :: Lens' CreateKeyPair Text

A unique name for the key pair.

Constraints: Up to 255 ASCII characters

Destructuring the Response

data CreateKeyPairResponse

Describes a key pair.

See: newCreateKeyPairResponse smart constructor.

Constructors

CreateKeyPairResponse'

Fields

keyPairId :: Maybe Text

The ID of the key pair.

tags :: Maybe [Tag]

Any tags applied to the key pair.

httpStatus :: Int

The response's http status code.

keyName :: Text

The name of the key pair.

keyFingerprint :: Text

The SHA-1 digest of the DER encoded private key.

keyMaterial :: Sensitive Text

An unencrypted PEM encoded RSA or ED25519 private key.

newCreateKeyPairResponse

Arguments

:: Int
$sel:httpStatus:CreateKeyPairResponse'

-> Text
$sel:keyName:CreateKeyPair'

-> Text
$sel:keyFingerprint:CreateKeyPairResponse'

-> Text
$sel:keyMaterial:CreateKeyPairResponse'

-> CreateKeyPairResponse

Create a value of CreateKeyPairResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:keyPairId:CreateKeyPairResponse', createKeyPairResponse_keyPairId - The ID of the key pair.

$sel:tags:CreateKeyPairResponse', createKeyPairResponse_tags - Any tags applied to the key pair.

$sel:httpStatus:CreateKeyPairResponse', createKeyPairResponse_httpStatus - The response's http status code.

$sel:keyName:CreateKeyPair', createKeyPairResponse_keyName - The name of the key pair.

$sel:keyFingerprint:CreateKeyPairResponse', createKeyPairResponse_keyFingerprint - The SHA-1 digest of the DER encoded private key.

$sel:keyMaterial:CreateKeyPairResponse', createKeyPairResponse_keyMaterial - An unencrypted PEM encoded RSA or ED25519 private key.

Response Lenses

createKeyPairResponse_keyPairId :: Lens' CreateKeyPairResponse (Maybe Text)

The ID of the key pair.

createKeyPairResponse_tags :: Lens' CreateKeyPairResponse (Maybe [Tag])

Any tags applied to the key pair.

createKeyPairResponse_httpStatus :: Lens' CreateKeyPairResponse Int

The response's http status code.

createKeyPairResponse_keyName :: Lens' CreateKeyPairResponse Text

The name of the key pair.

createKeyPairResponse_keyFingerprint :: Lens' CreateKeyPairResponse Text

The SHA-1 digest of the DER encoded private key.

createKeyPairResponse_keyMaterial :: Lens' CreateKeyPairResponse Text

An unencrypted PEM encoded RSA or ED25519 private key.