Amazonka.EC2.CreateLaunchTemplateVersion

Description

Creates a new version for a launch template. You can specify an existing version of launch template from which to base the new version.

Launch template versions are numbered in the order in which they are created. You cannot specify, change, or replace the numbering of launch template versions.

For more information, see Managing launch template versionsin the Amazon Elastic Compute Cloud User Guide.

Creating a Request

data CreateLaunchTemplateVersion

See: newCreateLaunchTemplateVersion smart constructor.

Constructors

CreateLaunchTemplateVersion'

Fields

launchTemplateName :: Maybe Text

The name of the launch template. You must specify either the launch template ID or launch template name in the request.

clientToken :: Maybe Text

Unique, case-sensitive identifier you provide to ensure the idempotency of the request. For more information, see Ensuring Idempotency.

Constraint: Maximum 128 ASCII characters.

launchTemplateId :: Maybe Text

The ID of the launch template. You must specify either the launch template ID or launch template name in the request.

versionDescription :: Maybe Text

A description for the version of the launch template.

sourceVersion :: Maybe Text

The version number of the launch template version on which to base the new version. The new version inherits the same launch parameters as the source version, except for parameters that you specify in LaunchTemplateData. Snapshots applied to the block device mapping are ignored when creating a new version unless they are explicitly included.

dryRun :: Maybe Bool

Checks whether you have the required permissions for the action, without actually making the request, and provides an error response. If you have the required permissions, the error response is DryRunOperation. Otherwise, it is UnauthorizedOperation.

launchTemplateData :: RequestLaunchTemplateData

The information for the launch template.

newCreateLaunchTemplateVersion

Arguments

:: RequestLaunchTemplateData
$sel:launchTemplateData:CreateLaunchTemplateVersion'

-> CreateLaunchTemplateVersion

Create a value of CreateLaunchTemplateVersion with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:launchTemplateName:CreateLaunchTemplateVersion', createLaunchTemplateVersion_launchTemplateName - The name of the launch template. You must specify either the launch template ID or launch template name in the request.

$sel:clientToken:CreateLaunchTemplateVersion', createLaunchTemplateVersion_clientToken - Unique, case-sensitive identifier you provide to ensure the idempotency of the request. For more information, see Ensuring Idempotency.

Constraint: Maximum 128 ASCII characters.

$sel:launchTemplateId:CreateLaunchTemplateVersion', createLaunchTemplateVersion_launchTemplateId - The ID of the launch template. You must specify either the launch template ID or launch template name in the request.

$sel:versionDescription:CreateLaunchTemplateVersion', createLaunchTemplateVersion_versionDescription - A description for the version of the launch template.

$sel:sourceVersion:CreateLaunchTemplateVersion', createLaunchTemplateVersion_sourceVersion - The version number of the launch template version on which to base the new version. The new version inherits the same launch parameters as the source version, except for parameters that you specify in LaunchTemplateData. Snapshots applied to the block device mapping are ignored when creating a new version unless they are explicitly included.

$sel:dryRun:CreateLaunchTemplateVersion', createLaunchTemplateVersion_dryRun - Checks whether you have the required permissions for the action, without actually making the request, and provides an error response. If you have the required permissions, the error response is DryRunOperation. Otherwise, it is UnauthorizedOperation.

$sel:launchTemplateData:CreateLaunchTemplateVersion', createLaunchTemplateVersion_launchTemplateData - The information for the launch template.

Request Lenses

createLaunchTemplateVersion_launchTemplateName :: Lens' CreateLaunchTemplateVersion (Maybe Text)

The name of the launch template. You must specify either the launch template ID or launch template name in the request.

createLaunchTemplateVersion_clientToken :: Lens' CreateLaunchTemplateVersion (Maybe Text)

Unique, case-sensitive identifier you provide to ensure the idempotency of the request. For more information, see Ensuring Idempotency.

Constraint: Maximum 128 ASCII characters.

createLaunchTemplateVersion_launchTemplateId :: Lens' CreateLaunchTemplateVersion (Maybe Text)

The ID of the launch template. You must specify either the launch template ID or launch template name in the request.

createLaunchTemplateVersion_versionDescription :: Lens' CreateLaunchTemplateVersion (Maybe Text)

A description for the version of the launch template.

createLaunchTemplateVersion_sourceVersion :: Lens' CreateLaunchTemplateVersion (Maybe Text)

The version number of the launch template version on which to base the new version. The new version inherits the same launch parameters as the source version, except for parameters that you specify in LaunchTemplateData. Snapshots applied to the block device mapping are ignored when creating a new version unless they are explicitly included.

createLaunchTemplateVersion_dryRun :: Lens' CreateLaunchTemplateVersion (Maybe Bool)

Checks whether you have the required permissions for the action, without actually making the request, and provides an error response. If you have the required permissions, the error response is DryRunOperation. Otherwise, it is UnauthorizedOperation.

createLaunchTemplateVersion_launchTemplateData :: Lens' CreateLaunchTemplateVersion RequestLaunchTemplateData

The information for the launch template.

Destructuring the Response

data CreateLaunchTemplateVersionResponse

See: newCreateLaunchTemplateVersionResponse smart constructor.

Constructors

CreateLaunchTemplateVersionResponse'

Fields

launchTemplateVersion :: Maybe LaunchTemplateVersion

Information about the launch template version.

warning :: Maybe ValidationWarning

If the new version of the launch template contains parameters or parameter combinations that are not valid, an error code and an error message are returned for each issue that's found.

httpStatus :: Int

The response's http status code.

newCreateLaunchTemplateVersionResponse

Arguments

:: Int
$sel:httpStatus:CreateLaunchTemplateVersionResponse'

-> CreateLaunchTemplateVersionResponse

Create a value of CreateLaunchTemplateVersionResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:launchTemplateVersion:CreateLaunchTemplateVersionResponse', createLaunchTemplateVersionResponse_launchTemplateVersion - Information about the launch template version.

$sel:warning:CreateLaunchTemplateVersionResponse', createLaunchTemplateVersionResponse_warning - If the new version of the launch template contains parameters or parameter combinations that are not valid, an error code and an error message are returned for each issue that's found.

$sel:httpStatus:CreateLaunchTemplateVersionResponse', createLaunchTemplateVersionResponse_httpStatus - The response's http status code.

Response Lenses

createLaunchTemplateVersionResponse_launchTemplateVersion :: Lens' CreateLaunchTemplateVersionResponse (Maybe LaunchTemplateVersion)

Information about the launch template version.

createLaunchTemplateVersionResponse_warning :: Lens' CreateLaunchTemplateVersionResponse (Maybe ValidationWarning)

If the new version of the launch template contains parameters or parameter combinations that are not valid, an error code and an error message are returned for each issue that's found.

createLaunchTemplateVersionResponse_httpStatus :: Lens' CreateLaunchTemplateVersionResponse Int

The response's http status code.