Amazonka.EC2.CreateNetworkAclEntry

Description

Creates an entry (a rule) in a network ACL with the specified rule number. Each network ACL has a set of numbered ingress rules and a separate set of numbered egress rules. When determining whether a packet should be allowed in or out of a subnet associated with the ACL, we process the entries in the ACL according to the rule numbers, in ascending order. Each network ACL has a set of ingress rules and a separate set of egress rules.

We recommend that you leave room between the rule numbers (for example, 100, 110, 120, ...), and not number them one right after the other (for example, 101, 102, 103, ...). This makes it easier to add a rule between existing ones without having to renumber the rules.

After you add an entry, you can't modify it; you must either replace it, or create an entry and delete the old one.

For more information about network ACLs, see Network ACLs in the Amazon Virtual Private Cloud User Guide.

Creating a Request

data CreateNetworkAclEntry

See: newCreateNetworkAclEntry smart constructor.

Constructors

CreateNetworkAclEntry'

Fields

ipv6CidrBlock :: Maybe Text

The IPv6 network range to allow or deny, in CIDR notation (for example 2001:db8:1234:1a00::/64).

icmpTypeCode :: Maybe IcmpTypeCode

ICMP protocol: The ICMP or ICMPv6 type and code. Required if specifying protocol 1 (ICMP) or protocol 58 (ICMPv6) with an IPv6 CIDR block.

portRange :: Maybe PortRange

TCP or UDP protocols: The range of ports the rule applies to. Required if specifying protocol 6 (TCP) or 17 (UDP).

cidrBlock :: Maybe Text

The IPv4 network range to allow or deny, in CIDR notation (for example 172.16.0.0/24). We modify the specified CIDR block to its canonical form; for example, if you specify 100.68.0.18/18, we modify it to 100.68.0.0/18.

dryRun :: Maybe Bool

Checks whether you have the required permissions for the action, without actually making the request, and provides an error response. If you have the required permissions, the error response is DryRunOperation. Otherwise, it is UnauthorizedOperation.

egress :: Bool

Indicates whether this is an egress rule (rule is applied to traffic leaving the subnet).

networkAclId :: Text

The ID of the network ACL.

protocol :: Text

The protocol number. A value of "-1" means all protocols. If you specify "-1" or a protocol number other than "6" (TCP), "17" (UDP), or "1" (ICMP), traffic on all ports is allowed, regardless of any ports or ICMP types or codes that you specify. If you specify protocol "58" (ICMPv6) and specify an IPv4 CIDR block, traffic for all ICMP types and codes allowed, regardless of any that you specify. If you specify protocol "58" (ICMPv6) and specify an IPv6 CIDR block, you must specify an ICMP type and code.

ruleAction :: RuleAction

Indicates whether to allow or deny the traffic that matches the rule.

ruleNumber :: Int

The rule number for the entry (for example, 100). ACL entries are processed in ascending order by rule number.

Constraints: Positive integer from 1 to 32766. The range 32767 to 65535 is reserved for internal use.

newCreateNetworkAclEntry

Arguments

:: Bool
$sel:egress:CreateNetworkAclEntry'

-> Text
$sel:networkAclId:CreateNetworkAclEntry'

-> Text
$sel:protocol:CreateNetworkAclEntry'

-> RuleAction
$sel:ruleAction:CreateNetworkAclEntry'

-> Int
$sel:ruleNumber:CreateNetworkAclEntry'

-> CreateNetworkAclEntry

Create a value of CreateNetworkAclEntry with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:ipv6CidrBlock:CreateNetworkAclEntry', createNetworkAclEntry_ipv6CidrBlock - The IPv6 network range to allow or deny, in CIDR notation (for example 2001:db8:1234:1a00::/64).

$sel:icmpTypeCode:CreateNetworkAclEntry', createNetworkAclEntry_icmpTypeCode - ICMP protocol: The ICMP or ICMPv6 type and code. Required if specifying protocol 1 (ICMP) or protocol 58 (ICMPv6) with an IPv6 CIDR block.

$sel:portRange:CreateNetworkAclEntry', createNetworkAclEntry_portRange - TCP or UDP protocols: The range of ports the rule applies to. Required if specifying protocol 6 (TCP) or 17 (UDP).

$sel:cidrBlock:CreateNetworkAclEntry', createNetworkAclEntry_cidrBlock - The IPv4 network range to allow or deny, in CIDR notation (for example 172.16.0.0/24). We modify the specified CIDR block to its canonical form; for example, if you specify 100.68.0.18/18, we modify it to 100.68.0.0/18.

$sel:dryRun:CreateNetworkAclEntry', createNetworkAclEntry_dryRun - Checks whether you have the required permissions for the action, without actually making the request, and provides an error response. If you have the required permissions, the error response is DryRunOperation. Otherwise, it is UnauthorizedOperation.

$sel:egress:CreateNetworkAclEntry', createNetworkAclEntry_egress - Indicates whether this is an egress rule (rule is applied to traffic leaving the subnet).

$sel:networkAclId:CreateNetworkAclEntry', createNetworkAclEntry_networkAclId - The ID of the network ACL.

$sel:protocol:CreateNetworkAclEntry', createNetworkAclEntry_protocol - The protocol number. A value of "-1" means all protocols. If you specify "-1" or a protocol number other than "6" (TCP), "17" (UDP), or "1" (ICMP), traffic on all ports is allowed, regardless of any ports or ICMP types or codes that you specify. If you specify protocol "58" (ICMPv6) and specify an IPv4 CIDR block, traffic for all ICMP types and codes allowed, regardless of any that you specify. If you specify protocol "58" (ICMPv6) and specify an IPv6 CIDR block, you must specify an ICMP type and code.

$sel:ruleAction:CreateNetworkAclEntry', createNetworkAclEntry_ruleAction - Indicates whether to allow or deny the traffic that matches the rule.

$sel:ruleNumber:CreateNetworkAclEntry', createNetworkAclEntry_ruleNumber - The rule number for the entry (for example, 100). ACL entries are processed in ascending order by rule number.

Constraints: Positive integer from 1 to 32766. The range 32767 to 65535 is reserved for internal use.

Request Lenses

createNetworkAclEntry_ipv6CidrBlock :: Lens' CreateNetworkAclEntry (Maybe Text)

The IPv6 network range to allow or deny, in CIDR notation (for example 2001:db8:1234:1a00::/64).

createNetworkAclEntry_icmpTypeCode :: Lens' CreateNetworkAclEntry (Maybe IcmpTypeCode)

ICMP protocol: The ICMP or ICMPv6 type and code. Required if specifying protocol 1 (ICMP) or protocol 58 (ICMPv6) with an IPv6 CIDR block.

createNetworkAclEntry_portRange :: Lens' CreateNetworkAclEntry (Maybe PortRange)

TCP or UDP protocols: The range of ports the rule applies to. Required if specifying protocol 6 (TCP) or 17 (UDP).

createNetworkAclEntry_cidrBlock :: Lens' CreateNetworkAclEntry (Maybe Text)

The IPv4 network range to allow or deny, in CIDR notation (for example 172.16.0.0/24). We modify the specified CIDR block to its canonical form; for example, if you specify 100.68.0.18/18, we modify it to 100.68.0.0/18.

createNetworkAclEntry_dryRun :: Lens' CreateNetworkAclEntry (Maybe Bool)

Checks whether you have the required permissions for the action, without actually making the request, and provides an error response. If you have the required permissions, the error response is DryRunOperation. Otherwise, it is UnauthorizedOperation.

createNetworkAclEntry_egress :: Lens' CreateNetworkAclEntry Bool

Indicates whether this is an egress rule (rule is applied to traffic leaving the subnet).

createNetworkAclEntry_networkAclId :: Lens' CreateNetworkAclEntry Text

The ID of the network ACL.

createNetworkAclEntry_protocol :: Lens' CreateNetworkAclEntry Text

The protocol number. A value of "-1" means all protocols. If you specify "-1" or a protocol number other than "6" (TCP), "17" (UDP), or "1" (ICMP), traffic on all ports is allowed, regardless of any ports or ICMP types or codes that you specify. If you specify protocol "58" (ICMPv6) and specify an IPv4 CIDR block, traffic for all ICMP types and codes allowed, regardless of any that you specify. If you specify protocol "58" (ICMPv6) and specify an IPv6 CIDR block, you must specify an ICMP type and code.

createNetworkAclEntry_ruleAction :: Lens' CreateNetworkAclEntry RuleAction

Indicates whether to allow or deny the traffic that matches the rule.

createNetworkAclEntry_ruleNumber :: Lens' CreateNetworkAclEntry Int

The rule number for the entry (for example, 100). ACL entries are processed in ascending order by rule number.

Constraints: Positive integer from 1 to 32766. The range 32767 to 65535 is reserved for internal use.

Destructuring the Response

data CreateNetworkAclEntryResponse

See: newCreateNetworkAclEntryResponse smart constructor.

Constructors

CreateNetworkAclEntryResponse'

newCreateNetworkAclEntryResponse :: CreateNetworkAclEntryResponse

Create a value of CreateNetworkAclEntryResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.