Amazonka.EC2.CreateNetworkInterface

Description

Creates a network interface in the specified subnet.

For more information about network interfaces, see Elastic Network Interfaces in the Amazon Virtual Private Cloud User Guide.

Creating a Request

data CreateNetworkInterface

Contains the parameters for CreateNetworkInterface.

See: newCreateNetworkInterface smart constructor.

Constructors

CreateNetworkInterface'

Fields

groups :: Maybe [Text]

The IDs of one or more security groups.

privateIpAddresses :: Maybe [PrivateIpAddressSpecification]

One or more private IPv4 addresses.

clientToken :: Maybe Text

Unique, case-sensitive identifier that you provide to ensure the idempotency of the request. For more information, see Ensuring Idempotency.

ipv4Prefixes :: Maybe [Ipv4PrefixSpecificationRequest]

One or more IPv4 prefixes assigned to the network interface. You cannot use this option if you use the Ipv4PrefixCount option.

interfaceType :: Maybe NetworkInterfaceCreationType

Indicates the type of network interface. To create an Elastic Fabric Adapter (EFA), specify efa. For more information, see Elastic Fabric Adapter in the Amazon Elastic Compute Cloud User Guide. To create a trunk network interface, specify efa. For more information, see Network interface trunking in the Amazon Elastic Compute Cloud User Guide.

ipv4PrefixCount :: Maybe Int

The number of IPv4 prefixes that Amazon Web Services automatically assigns to the network interface. You cannot use this option if you use the Ipv4 Prefixes option.

tagSpecifications :: Maybe [TagSpecification]

The tags to apply to the new network interface.

ipv6AddressCount :: Maybe Int

The number of IPv6 addresses to assign to a network interface. Amazon EC2 automatically selects the IPv6 addresses from the subnet range. You can't use this option if specifying specific IPv6 addresses. If your subnet has the AssignIpv6AddressOnCreation attribute set to true, you can specify 0 to override this setting.

ipv6Prefixes :: Maybe [Ipv6PrefixSpecificationRequest]

One or more IPv6 prefixes assigned to the network interface. You cannot use this option if you use the Ipv6PrefixCount option.

privateIpAddress :: Maybe Text

The primary private IPv4 address of the network interface. If you don't specify an IPv4 address, Amazon EC2 selects one for you from the subnet's IPv4 CIDR range. If you specify an IP address, you cannot indicate any IP addresses specified in privateIpAddresses as primary (only one IP address can be designated as primary).

ipv6PrefixCount :: Maybe Int

The number of IPv6 prefixes that Amazon Web Services automatically assigns to the network interface. You cannot use this option if you use the Ipv6Prefixes option.

secondaryPrivateIpAddressCount :: Maybe Int

The number of secondary private IPv4 addresses to assign to a network interface. When you specify a number of secondary IPv4 addresses, Amazon EC2 selects these IP addresses within the subnet's IPv4 CIDR range. You can't specify this option and specify more than one private IP address using privateIpAddresses.

The number of IP addresses you can assign to a network interface varies by instance type. For more information, see IP Addresses Per ENI Per Instance Type in the Amazon Virtual Private Cloud User Guide.

description :: Maybe Text

A description for the network interface.

dryRun :: Maybe Bool

Checks whether you have the required permissions for the action, without actually making the request, and provides an error response. If you have the required permissions, the error response is DryRunOperation. Otherwise, it is UnauthorizedOperation.

ipv6Addresses :: Maybe [InstanceIpv6Address]

One or more specific IPv6 addresses from the IPv6 CIDR block range of your subnet. You can't use this option if you're specifying a number of IPv6 addresses.

subnetId :: Text

The ID of the subnet to associate with the network interface.

newCreateNetworkInterface

Arguments

:: Text
$sel:subnetId:CreateNetworkInterface'

-> CreateNetworkInterface

Create a value of CreateNetworkInterface with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:groups:CreateNetworkInterface', createNetworkInterface_groups - The IDs of one or more security groups.

$sel:privateIpAddresses:CreateNetworkInterface', createNetworkInterface_privateIpAddresses - One or more private IPv4 addresses.

$sel:clientToken:CreateNetworkInterface', createNetworkInterface_clientToken - Unique, case-sensitive identifier that you provide to ensure the idempotency of the request. For more information, see Ensuring Idempotency.

$sel:ipv4Prefixes:CreateNetworkInterface', createNetworkInterface_ipv4Prefixes - One or more IPv4 prefixes assigned to the network interface. You cannot use this option if you use the Ipv4PrefixCount option.

$sel:interfaceType:CreateNetworkInterface', createNetworkInterface_interfaceType - Indicates the type of network interface. To create an Elastic Fabric Adapter (EFA), specify efa. For more information, see Elastic Fabric Adapter in the Amazon Elastic Compute Cloud User Guide. To create a trunk network interface, specify efa. For more information, see Network interface trunking in the Amazon Elastic Compute Cloud User Guide.

$sel:ipv4PrefixCount:CreateNetworkInterface', createNetworkInterface_ipv4PrefixCount - The number of IPv4 prefixes that Amazon Web Services automatically assigns to the network interface. You cannot use this option if you use the Ipv4 Prefixes option.

$sel:tagSpecifications:CreateNetworkInterface', createNetworkInterface_tagSpecifications - The tags to apply to the new network interface.

$sel:ipv6AddressCount:CreateNetworkInterface', createNetworkInterface_ipv6AddressCount - The number of IPv6 addresses to assign to a network interface. Amazon EC2 automatically selects the IPv6 addresses from the subnet range. You can't use this option if specifying specific IPv6 addresses. If your subnet has the AssignIpv6AddressOnCreation attribute set to true, you can specify 0 to override this setting.

$sel:ipv6Prefixes:CreateNetworkInterface', createNetworkInterface_ipv6Prefixes - One or more IPv6 prefixes assigned to the network interface. You cannot use this option if you use the Ipv6PrefixCount option.

$sel:privateIpAddress:CreateNetworkInterface', createNetworkInterface_privateIpAddress - The primary private IPv4 address of the network interface. If you don't specify an IPv4 address, Amazon EC2 selects one for you from the subnet's IPv4 CIDR range. If you specify an IP address, you cannot indicate any IP addresses specified in privateIpAddresses as primary (only one IP address can be designated as primary).

$sel:ipv6PrefixCount:CreateNetworkInterface', createNetworkInterface_ipv6PrefixCount - The number of IPv6 prefixes that Amazon Web Services automatically assigns to the network interface. You cannot use this option if you use the Ipv6Prefixes option.

$sel:secondaryPrivateIpAddressCount:CreateNetworkInterface', createNetworkInterface_secondaryPrivateIpAddressCount - The number of secondary private IPv4 addresses to assign to a network interface. When you specify a number of secondary IPv4 addresses, Amazon EC2 selects these IP addresses within the subnet's IPv4 CIDR range. You can't specify this option and specify more than one private IP address using privateIpAddresses.

The number of IP addresses you can assign to a network interface varies by instance type. For more information, see IP Addresses Per ENI Per Instance Type in the Amazon Virtual Private Cloud User Guide.

$sel:description:CreateNetworkInterface', createNetworkInterface_description - A description for the network interface.

$sel:dryRun:CreateNetworkInterface', createNetworkInterface_dryRun - Checks whether you have the required permissions for the action, without actually making the request, and provides an error response. If you have the required permissions, the error response is DryRunOperation. Otherwise, it is UnauthorizedOperation.

$sel:ipv6Addresses:CreateNetworkInterface', createNetworkInterface_ipv6Addresses - One or more specific IPv6 addresses from the IPv6 CIDR block range of your subnet. You can't use this option if you're specifying a number of IPv6 addresses.

$sel:subnetId:CreateNetworkInterface', createNetworkInterface_subnetId - The ID of the subnet to associate with the network interface.

Request Lenses

createNetworkInterface_groups :: Lens' CreateNetworkInterface (Maybe [Text])

The IDs of one or more security groups.

createNetworkInterface_privateIpAddresses :: Lens' CreateNetworkInterface (Maybe [PrivateIpAddressSpecification])

One or more private IPv4 addresses.

createNetworkInterface_clientToken :: Lens' CreateNetworkInterface (Maybe Text)

Unique, case-sensitive identifier that you provide to ensure the idempotency of the request. For more information, see Ensuring Idempotency.

createNetworkInterface_ipv4Prefixes :: Lens' CreateNetworkInterface (Maybe [Ipv4PrefixSpecificationRequest])

One or more IPv4 prefixes assigned to the network interface. You cannot use this option if you use the Ipv4PrefixCount option.

createNetworkInterface_interfaceType :: Lens' CreateNetworkInterface (Maybe NetworkInterfaceCreationType)

Indicates the type of network interface. To create an Elastic Fabric Adapter (EFA), specify efa. For more information, see Elastic Fabric Adapter in the Amazon Elastic Compute Cloud User Guide. To create a trunk network interface, specify efa. For more information, see Network interface trunking in the Amazon Elastic Compute Cloud User Guide.

createNetworkInterface_ipv4PrefixCount :: Lens' CreateNetworkInterface (Maybe Int)

The number of IPv4 prefixes that Amazon Web Services automatically assigns to the network interface. You cannot use this option if you use the Ipv4 Prefixes option.

createNetworkInterface_tagSpecifications :: Lens' CreateNetworkInterface (Maybe [TagSpecification])

The tags to apply to the new network interface.

createNetworkInterface_ipv6AddressCount :: Lens' CreateNetworkInterface (Maybe Int)

The number of IPv6 addresses to assign to a network interface. Amazon EC2 automatically selects the IPv6 addresses from the subnet range. You can't use this option if specifying specific IPv6 addresses. If your subnet has the AssignIpv6AddressOnCreation attribute set to true, you can specify 0 to override this setting.

createNetworkInterface_ipv6Prefixes :: Lens' CreateNetworkInterface (Maybe [Ipv6PrefixSpecificationRequest])

One or more IPv6 prefixes assigned to the network interface. You cannot use this option if you use the Ipv6PrefixCount option.

createNetworkInterface_privateIpAddress :: Lens' CreateNetworkInterface (Maybe Text)

The primary private IPv4 address of the network interface. If you don't specify an IPv4 address, Amazon EC2 selects one for you from the subnet's IPv4 CIDR range. If you specify an IP address, you cannot indicate any IP addresses specified in privateIpAddresses as primary (only one IP address can be designated as primary).

createNetworkInterface_ipv6PrefixCount :: Lens' CreateNetworkInterface (Maybe Int)

The number of IPv6 prefixes that Amazon Web Services automatically assigns to the network interface. You cannot use this option if you use the Ipv6Prefixes option.

createNetworkInterface_secondaryPrivateIpAddressCount :: Lens' CreateNetworkInterface (Maybe Int)

The number of secondary private IPv4 addresses to assign to a network interface. When you specify a number of secondary IPv4 addresses, Amazon EC2 selects these IP addresses within the subnet's IPv4 CIDR range. You can't specify this option and specify more than one private IP address using privateIpAddresses.

The number of IP addresses you can assign to a network interface varies by instance type. For more information, see IP Addresses Per ENI Per Instance Type in the Amazon Virtual Private Cloud User Guide.

createNetworkInterface_description :: Lens' CreateNetworkInterface (Maybe Text)

A description for the network interface.

createNetworkInterface_dryRun :: Lens' CreateNetworkInterface (Maybe Bool)

Checks whether you have the required permissions for the action, without actually making the request, and provides an error response. If you have the required permissions, the error response is DryRunOperation. Otherwise, it is UnauthorizedOperation.

createNetworkInterface_ipv6Addresses :: Lens' CreateNetworkInterface (Maybe [InstanceIpv6Address])

One or more specific IPv6 addresses from the IPv6 CIDR block range of your subnet. You can't use this option if you're specifying a number of IPv6 addresses.

createNetworkInterface_subnetId :: Lens' CreateNetworkInterface Text

The ID of the subnet to associate with the network interface.

Destructuring the Response

data CreateNetworkInterfaceResponse

Contains the output of CreateNetworkInterface.

See: newCreateNetworkInterfaceResponse smart constructor.

Constructors

CreateNetworkInterfaceResponse'

Fields

clientToken :: Maybe Text

The token to use to retrieve the next page of results. This value is null when there are no more results to return.

networkInterface :: Maybe NetworkInterface

Information about the network interface.

httpStatus :: Int

The response's http status code.

newCreateNetworkInterfaceResponse

Arguments

:: Int
$sel:httpStatus:CreateNetworkInterfaceResponse'

-> CreateNetworkInterfaceResponse

Create a value of CreateNetworkInterfaceResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:clientToken:CreateNetworkInterface', createNetworkInterfaceResponse_clientToken - The token to use to retrieve the next page of results. This value is null when there are no more results to return.

$sel:networkInterface:CreateNetworkInterfaceResponse', createNetworkInterfaceResponse_networkInterface - Information about the network interface.

$sel:httpStatus:CreateNetworkInterfaceResponse', createNetworkInterfaceResponse_httpStatus - The response's http status code.

Response Lenses

createNetworkInterfaceResponse_clientToken :: Lens' CreateNetworkInterfaceResponse (Maybe Text)

The token to use to retrieve the next page of results. This value is null when there are no more results to return.

createNetworkInterfaceResponse_networkInterface :: Lens' CreateNetworkInterfaceResponse (Maybe NetworkInterface)

Information about the network interface.

createNetworkInterfaceResponse_httpStatus :: Lens' CreateNetworkInterfaceResponse Int

The response's http status code.