Amazonka.EC2.CreateSecurityGroup

Description

Creates a security group.

A security group acts as a virtual firewall for your instance to control inbound and outbound traffic. For more information, see Amazon EC2 security groups in the Amazon Elastic Compute Cloud User Guide and Security groups for your VPC in the Amazon Virtual Private Cloud User Guide.

When you create a security group, you specify a friendly name of your choice. You can have a security group for use in EC2-Classic with the same name as a security group for use in a VPC. However, you can't have two security groups for use in EC2-Classic with the same name or two security groups for use in a VPC with the same name.

You have a default security group for use in EC2-Classic and a default security group for use in your VPC. If you don't specify a security group when you launch an instance, the instance is launched into the appropriate default security group. A default security group includes a default rule that grants instances unrestricted network access to each other.

You can add or remove rules from your security groups using AuthorizeSecurityGroupIngress, AuthorizeSecurityGroupEgress, RevokeSecurityGroupIngress, and RevokeSecurityGroupEgress.

For more information about VPC security group limits, see Amazon VPC Limits.

Creating a Request

data CreateSecurityGroup

See: newCreateSecurityGroup smart constructor.

Constructors

CreateSecurityGroup'

Fields

vpcId :: Maybe Text

EC2-VPC

The ID of the VPC. Required for EC2-VPC.

tagSpecifications :: Maybe [TagSpecification]

The tags to assign to the security group.

dryRun :: Maybe Bool

Checks whether you have the required permissions for the action, without actually making the request, and provides an error response. If you have the required permissions, the error response is DryRunOperation. Otherwise, it is UnauthorizedOperation.

description :: Text

A description for the security group. This is informational only.

Constraints: Up to 255 characters in length

Constraints for EC2-Classic: ASCII characters

Constraints for EC2-VPC: a-z, A-Z, 0-9, spaces, and ._-:/()#,@[]+=&;{}!$*

groupName :: Text

The name of the security group.

Constraints: Up to 255 characters in length. Cannot start with sg-.

Constraints for EC2-Classic: ASCII characters

Constraints for EC2-VPC: a-z, A-Z, 0-9, spaces, and ._-:/()#,@[]+=&;{}!$*

newCreateSecurityGroup

Arguments

:: Text
$sel:description:CreateSecurityGroup'

-> Text
$sel:groupName:CreateSecurityGroup'

-> CreateSecurityGroup

Create a value of CreateSecurityGroup with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:vpcId:CreateSecurityGroup', createSecurityGroup_vpcId - [EC2-VPC] The ID of the VPC. Required for EC2-VPC.

$sel:tagSpecifications:CreateSecurityGroup', createSecurityGroup_tagSpecifications - The tags to assign to the security group.

$sel:dryRun:CreateSecurityGroup', createSecurityGroup_dryRun - Checks whether you have the required permissions for the action, without actually making the request, and provides an error response. If you have the required permissions, the error response is DryRunOperation. Otherwise, it is UnauthorizedOperation.

$sel:description:CreateSecurityGroup', createSecurityGroup_description - A description for the security group. This is informational only.

Constraints: Up to 255 characters in length

Constraints for EC2-Classic: ASCII characters

Constraints for EC2-VPC: a-z, A-Z, 0-9, spaces, and ._-:/()#,@[]+=&;{}!$*

$sel:groupName:CreateSecurityGroup', createSecurityGroup_groupName - The name of the security group.

Constraints: Up to 255 characters in length. Cannot start with sg-.

Constraints for EC2-Classic: ASCII characters

Constraints for EC2-VPC: a-z, A-Z, 0-9, spaces, and ._-:/()#,@[]+=&;{}!$*

Request Lenses

createSecurityGroup_vpcId :: Lens' CreateSecurityGroup (Maybe Text)

EC2-VPC

The ID of the VPC. Required for EC2-VPC.

createSecurityGroup_tagSpecifications :: Lens' CreateSecurityGroup (Maybe [TagSpecification])

The tags to assign to the security group.

createSecurityGroup_dryRun :: Lens' CreateSecurityGroup (Maybe Bool)

Checks whether you have the required permissions for the action, without actually making the request, and provides an error response. If you have the required permissions, the error response is DryRunOperation. Otherwise, it is UnauthorizedOperation.

createSecurityGroup_description :: Lens' CreateSecurityGroup Text

A description for the security group. This is informational only.

Constraints: Up to 255 characters in length

Constraints for EC2-Classic: ASCII characters

Constraints for EC2-VPC: a-z, A-Z, 0-9, spaces, and ._-:/()#,@[]+=&;{}!$*

createSecurityGroup_groupName :: Lens' CreateSecurityGroup Text

The name of the security group.

Constraints: Up to 255 characters in length. Cannot start with sg-.

Constraints for EC2-Classic: ASCII characters

Constraints for EC2-VPC: a-z, A-Z, 0-9, spaces, and ._-:/()#,@[]+=&;{}!$*

Destructuring the Response

data CreateSecurityGroupResponse

See: newCreateSecurityGroupResponse smart constructor.

Constructors

CreateSecurityGroupResponse'

Fields

tags :: Maybe [Tag]

The tags assigned to the security group.

httpStatus :: Int

The response's http status code.

groupId :: Text

The ID of the security group.

newCreateSecurityGroupResponse

Arguments

:: Int
$sel:httpStatus:CreateSecurityGroupResponse'

-> Text
$sel:groupId:CreateSecurityGroupResponse'

-> CreateSecurityGroupResponse

Create a value of CreateSecurityGroupResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:tags:CreateSecurityGroupResponse', createSecurityGroupResponse_tags - The tags assigned to the security group.

$sel:httpStatus:CreateSecurityGroupResponse', createSecurityGroupResponse_httpStatus - The response's http status code.

$sel:groupId:CreateSecurityGroupResponse', createSecurityGroupResponse_groupId - The ID of the security group.

Response Lenses

createSecurityGroupResponse_tags :: Lens' CreateSecurityGroupResponse (Maybe [Tag])

The tags assigned to the security group.

createSecurityGroupResponse_httpStatus :: Lens' CreateSecurityGroupResponse Int

The response's http status code.

createSecurityGroupResponse_groupId :: Lens' CreateSecurityGroupResponse Text

The ID of the security group.