Amazonka.EC2.CreateTransitGatewayConnectPeer

Description

Creates a Connect peer for a specified transit gateway Connect attachment between a transit gateway and an appliance.

The peer address and transit gateway address must be the same IP address family (IPv4 or IPv6).

For more information, see Connect peers in the Transit Gateways Guide.

Creating a Request

data CreateTransitGatewayConnectPeer

See: newCreateTransitGatewayConnectPeer smart constructor.

Constructors

CreateTransitGatewayConnectPeer'

Fields

bgpOptions :: Maybe TransitGatewayConnectRequestBgpOptions

The BGP options for the Connect peer.

transitGatewayAddress :: Maybe Text

The peer IP address (GRE outer IP address) on the transit gateway side of the Connect peer, which must be specified from a transit gateway CIDR block. If not specified, Amazon automatically assigns the first available IP address from the transit gateway CIDR block.

tagSpecifications :: Maybe [TagSpecification]

The tags to apply to the Connect peer.

dryRun :: Maybe Bool

Checks whether you have the required permissions for the action, without actually making the request, and provides an error response. If you have the required permissions, the error response is DryRunOperation. Otherwise, it is UnauthorizedOperation.

transitGatewayAttachmentId :: Text

The ID of the Connect attachment.

peerAddress :: Text

The peer IP address (GRE outer IP address) on the appliance side of the Connect peer.

insideCidrBlocks :: [Text]

The range of inside IP addresses that are used for BGP peering. You must specify a size /29 IPv4 CIDR block from the 169.254.0.0/16 range. The first address from the range must be configured on the appliance as the BGP IP address. You can also optionally specify a size /125 IPv6 CIDR block from the fd00::/8 range.

newCreateTransitGatewayConnectPeer

Arguments

:: Text
$sel:transitGatewayAttachmentId:CreateTransitGatewayConnectPeer'

-> Text
$sel:peerAddress:CreateTransitGatewayConnectPeer'

-> CreateTransitGatewayConnectPeer

Create a value of CreateTransitGatewayConnectPeer with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:bgpOptions:CreateTransitGatewayConnectPeer', createTransitGatewayConnectPeer_bgpOptions - The BGP options for the Connect peer.

$sel:transitGatewayAddress:CreateTransitGatewayConnectPeer', createTransitGatewayConnectPeer_transitGatewayAddress - The peer IP address (GRE outer IP address) on the transit gateway side of the Connect peer, which must be specified from a transit gateway CIDR block. If not specified, Amazon automatically assigns the first available IP address from the transit gateway CIDR block.

$sel:tagSpecifications:CreateTransitGatewayConnectPeer', createTransitGatewayConnectPeer_tagSpecifications - The tags to apply to the Connect peer.

$sel:dryRun:CreateTransitGatewayConnectPeer', createTransitGatewayConnectPeer_dryRun - Checks whether you have the required permissions for the action, without actually making the request, and provides an error response. If you have the required permissions, the error response is DryRunOperation. Otherwise, it is UnauthorizedOperation.

$sel:transitGatewayAttachmentId:CreateTransitGatewayConnectPeer', createTransitGatewayConnectPeer_transitGatewayAttachmentId - The ID of the Connect attachment.

$sel:peerAddress:CreateTransitGatewayConnectPeer', createTransitGatewayConnectPeer_peerAddress - The peer IP address (GRE outer IP address) on the appliance side of the Connect peer.

$sel:insideCidrBlocks:CreateTransitGatewayConnectPeer', createTransitGatewayConnectPeer_insideCidrBlocks - The range of inside IP addresses that are used for BGP peering. You must specify a size /29 IPv4 CIDR block from the 169.254.0.0/16 range. The first address from the range must be configured on the appliance as the BGP IP address. You can also optionally specify a size /125 IPv6 CIDR block from the fd00::/8 range.

Request Lenses

createTransitGatewayConnectPeer_bgpOptions :: Lens' CreateTransitGatewayConnectPeer (Maybe TransitGatewayConnectRequestBgpOptions)

The BGP options for the Connect peer.

createTransitGatewayConnectPeer_transitGatewayAddress :: Lens' CreateTransitGatewayConnectPeer (Maybe Text)

The peer IP address (GRE outer IP address) on the transit gateway side of the Connect peer, which must be specified from a transit gateway CIDR block. If not specified, Amazon automatically assigns the first available IP address from the transit gateway CIDR block.

createTransitGatewayConnectPeer_tagSpecifications :: Lens' CreateTransitGatewayConnectPeer (Maybe [TagSpecification])

The tags to apply to the Connect peer.

createTransitGatewayConnectPeer_dryRun :: Lens' CreateTransitGatewayConnectPeer (Maybe Bool)

Checks whether you have the required permissions for the action, without actually making the request, and provides an error response. If you have the required permissions, the error response is DryRunOperation. Otherwise, it is UnauthorizedOperation.

createTransitGatewayConnectPeer_transitGatewayAttachmentId :: Lens' CreateTransitGatewayConnectPeer Text

The ID of the Connect attachment.

createTransitGatewayConnectPeer_peerAddress :: Lens' CreateTransitGatewayConnectPeer Text

The peer IP address (GRE outer IP address) on the appliance side of the Connect peer.

createTransitGatewayConnectPeer_insideCidrBlocks :: Lens' CreateTransitGatewayConnectPeer [Text]

The range of inside IP addresses that are used for BGP peering. You must specify a size /29 IPv4 CIDR block from the 169.254.0.0/16 range. The first address from the range must be configured on the appliance as the BGP IP address. You can also optionally specify a size /125 IPv6 CIDR block from the fd00::/8 range.

Destructuring the Response

data CreateTransitGatewayConnectPeerResponse

See: newCreateTransitGatewayConnectPeerResponse smart constructor.

Constructors

CreateTransitGatewayConnectPeerResponse'

Fields

transitGatewayConnectPeer :: Maybe TransitGatewayConnectPeer

Information about the Connect peer.

httpStatus :: Int

The response's http status code.

newCreateTransitGatewayConnectPeerResponse

Arguments

:: Int
$sel:httpStatus:CreateTransitGatewayConnectPeerResponse'

-> CreateTransitGatewayConnectPeerResponse

Create a value of CreateTransitGatewayConnectPeerResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:transitGatewayConnectPeer:CreateTransitGatewayConnectPeerResponse', createTransitGatewayConnectPeerResponse_transitGatewayConnectPeer - Information about the Connect peer.

$sel:httpStatus:CreateTransitGatewayConnectPeerResponse', createTransitGatewayConnectPeerResponse_httpStatus - The response's http status code.

Response Lenses

createTransitGatewayConnectPeerResponse_transitGatewayConnectPeer :: Lens' CreateTransitGatewayConnectPeerResponse (Maybe TransitGatewayConnectPeer)

Information about the Connect peer.

createTransitGatewayConnectPeerResponse_httpStatus :: Lens' CreateTransitGatewayConnectPeerResponse Int

The response's http status code.