Amazonka.EC2.CreateVolume

Description

Creates an EBS volume that can be attached to an instance in the same Availability Zone.

You can create a new empty volume or restore a volume from an EBS snapshot. Any Amazon Web Services Marketplace product codes from the snapshot are propagated to the volume.

You can create encrypted volumes. Encrypted volumes must be attached to instances that support Amazon EBS encryption. Volumes that are created from encrypted snapshots are also automatically encrypted. For more information, see Amazon EBS encryption in the Amazon Elastic Compute Cloud User Guide.

You can tag your volumes during creation. For more information, see Tag your Amazon EC2 resources in the Amazon Elastic Compute Cloud User Guide.

For more information, see Create an Amazon EBS volume in the Amazon Elastic Compute Cloud User Guide.

Creating a Request

data CreateVolume

See: newCreateVolume smart constructor.

Constructors

CreateVolume'

Fields

multiAttachEnabled :: Maybe Bool

Indicates whether to enable Amazon EBS Multi-Attach. If you enable Multi-Attach, you can attach the volume to up to 16 Instances built on the Nitro System in the same Availability Zone. This parameter is supported with io1 and io2 volumes only. For more information, see Amazon EBS Multi-Attach in the Amazon Elastic Compute Cloud User Guide.

clientToken :: Maybe Text

Unique, case-sensitive identifier that you provide to ensure the idempotency of the request. For more information, see Ensure Idempotency.

throughput :: Maybe Int

The throughput to provision for a volume, with a maximum of 1,000 MiB/s.

This parameter is valid only for gp3 volumes.

Valid Range: Minimum value of 125. Maximum value of 1000.

size :: Maybe Int

The size of the volume, in GiBs. You must specify either a snapshot ID or a volume size. If you specify a snapshot, the default is the snapshot size. You can specify a volume size that is equal to or larger than the snapshot size.

The following are the supported volumes sizes for each volume type:

gp2 and gp3: 1-16,384

io1 and io2: 4-16,384

st1 and sc1: 125-16,384

standard: 1-1,024

iops :: Maybe Int

The number of I/O operations per second (IOPS). For gp3, io1, and io2 volumes, this represents the number of IOPS that are provisioned for the volume. For gp2 volumes, this represents the baseline performance of the volume and the rate at which the volume accumulates I/O credits for bursting.

The following are the supported values for each volume type:

gp3: 3,000-16,000 IOPS

io1: 100-64,000 IOPS

io2: 100-64,000 IOPS

io1 and io2 volumes support up to 64,000 IOPS only on Instances built on the Nitro System. Other instance families support performance up to 32,000 IOPS.

This parameter is required for io1 and io2 volumes. The default for gp3 volumes is 3,000 IOPS. This parameter is not supported for gp2, st1, sc1, or standard volumes.

outpostArn :: Maybe Text

The Amazon Resource Name (ARN) of the Outpost.

encrypted :: Maybe Bool

Indicates whether the volume should be encrypted. The effect of setting the encryption state to true depends on the volume origin (new or from a snapshot), starting encryption state, ownership, and whether encryption by default is enabled. For more information, see Encryption by default in the Amazon Elastic Compute Cloud User Guide.

Encrypted Amazon EBS volumes must be attached to instances that support Amazon EBS encryption. For more information, see Supported instance types.

tagSpecifications :: Maybe [TagSpecification]

The tags to apply to the volume during creation.

kmsKeyId :: Maybe Text

The identifier of the Key Management Service (KMS) KMS key to use for Amazon EBS encryption. If this parameter is not specified, your KMS key for Amazon EBS is used. If KmsKeyId is specified, the encrypted state must be true.

You can specify the KMS key using any of the following:

Key ID. For example, 1234abcd-12ab-34cd-56ef-1234567890ab.

Key alias. For example, alias/ExampleAlias.

Key ARN. For example, arn:aws:kms:us-east-1:012345678910:key/1234abcd-12ab-34cd-56ef-1234567890ab.

Alias ARN. For example, arn:aws:kms:us-east-1:012345678910:alias/ExampleAlias.

Amazon Web Services authenticates the KMS key asynchronously. Therefore, if you specify an ID, alias, or ARN that is not valid, the action can appear to complete, but eventually fails.

volumeType :: Maybe VolumeType

The volume type. This parameter can be one of the following values:

General Purpose SSD: gp2 | gp3

Provisioned IOPS SSD: io1 | io2

Throughput Optimized HDD: st1

Cold HDD: sc1

Magnetic: standard

For more information, see Amazon EBS volume types in the Amazon Elastic Compute Cloud User Guide.

Default: gp2

dryRun :: Maybe Bool

Checks whether you have the required permissions for the action, without actually making the request, and provides an error response. If you have the required permissions, the error response is DryRunOperation. Otherwise, it is UnauthorizedOperation.

snapshotId :: Maybe Text

The snapshot from which to create the volume. You must specify either a snapshot ID or a volume size.

availabilityZone :: Text

The Availability Zone in which to create the volume.

newCreateVolume

Arguments

:: Text
$sel:availabilityZone:CreateVolume'

-> CreateVolume

Create a value of CreateVolume with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:multiAttachEnabled:CreateVolume', createVolume_multiAttachEnabled - Indicates whether to enable Amazon EBS Multi-Attach. If you enable Multi-Attach, you can attach the volume to up to 16 Instances built on the Nitro System in the same Availability Zone. This parameter is supported with io1 and io2 volumes only. For more information, see Amazon EBS Multi-Attach in the Amazon Elastic Compute Cloud User Guide.

$sel:clientToken:CreateVolume', createVolume_clientToken - Unique, case-sensitive identifier that you provide to ensure the idempotency of the request. For more information, see Ensure Idempotency.

$sel:throughput:CreateVolume', createVolume_throughput - The throughput to provision for a volume, with a maximum of 1,000 MiB/s.

This parameter is valid only for gp3 volumes.

Valid Range: Minimum value of 125. Maximum value of 1000.

$sel:size:CreateVolume', createVolume_size - The size of the volume, in GiBs. You must specify either a snapshot ID or a volume size. If you specify a snapshot, the default is the snapshot size. You can specify a volume size that is equal to or larger than the snapshot size.

The following are the supported volumes sizes for each volume type:

gp2 and gp3: 1-16,384

io1 and io2: 4-16,384

st1 and sc1: 125-16,384

standard: 1-1,024

$sel:iops:CreateVolume', createVolume_iops - The number of I/O operations per second (IOPS). For gp3, io1, and io2 volumes, this represents the number of IOPS that are provisioned for the volume. For gp2 volumes, this represents the baseline performance of the volume and the rate at which the volume accumulates I/O credits for bursting.

The following are the supported values for each volume type:

gp3: 3,000-16,000 IOPS

io1: 100-64,000 IOPS

io2: 100-64,000 IOPS

io1 and io2 volumes support up to 64,000 IOPS only on Instances built on the Nitro System. Other instance families support performance up to 32,000 IOPS.

This parameter is required for io1 and io2 volumes. The default for gp3 volumes is 3,000 IOPS. This parameter is not supported for gp2, st1, sc1, or standard volumes.

$sel:outpostArn:CreateVolume', createVolume_outpostArn - The Amazon Resource Name (ARN) of the Outpost.

$sel:encrypted:CreateVolume', createVolume_encrypted - Indicates whether the volume should be encrypted. The effect of setting the encryption state to true depends on the volume origin (new or from a snapshot), starting encryption state, ownership, and whether encryption by default is enabled. For more information, see Encryption by default in the Amazon Elastic Compute Cloud User Guide.

Encrypted Amazon EBS volumes must be attached to instances that support Amazon EBS encryption. For more information, see Supported instance types.

$sel:tagSpecifications:CreateVolume', createVolume_tagSpecifications - The tags to apply to the volume during creation.

$sel:kmsKeyId:CreateVolume', createVolume_kmsKeyId - The identifier of the Key Management Service (KMS) KMS key to use for Amazon EBS encryption. If this parameter is not specified, your KMS key for Amazon EBS is used. If KmsKeyId is specified, the encrypted state must be true.

You can specify the KMS key using any of the following:

Key ID. For example, 1234abcd-12ab-34cd-56ef-1234567890ab.

Key alias. For example, alias/ExampleAlias.

Key ARN. For example, arn:aws:kms:us-east-1:012345678910:key/1234abcd-12ab-34cd-56ef-1234567890ab.

Alias ARN. For example, arn:aws:kms:us-east-1:012345678910:alias/ExampleAlias.

Amazon Web Services authenticates the KMS key asynchronously. Therefore, if you specify an ID, alias, or ARN that is not valid, the action can appear to complete, but eventually fails.

$sel:volumeType:CreateVolume', createVolume_volumeType - The volume type. This parameter can be one of the following values:

General Purpose SSD: gp2 | gp3

Provisioned IOPS SSD: io1 | io2

Throughput Optimized HDD: st1

Cold HDD: sc1

Magnetic: standard

For more information, see Amazon EBS volume types in the Amazon Elastic Compute Cloud User Guide.

Default: gp2

$sel:dryRun:CreateVolume', createVolume_dryRun - Checks whether you have the required permissions for the action, without actually making the request, and provides an error response. If you have the required permissions, the error response is DryRunOperation. Otherwise, it is UnauthorizedOperation.

$sel:snapshotId:CreateVolume', createVolume_snapshotId - The snapshot from which to create the volume. You must specify either a snapshot ID or a volume size.

$sel:availabilityZone:CreateVolume', createVolume_availabilityZone - The Availability Zone in which to create the volume.

Request Lenses

createVolume_multiAttachEnabled :: Lens' CreateVolume (Maybe Bool)

Indicates whether to enable Amazon EBS Multi-Attach. If you enable Multi-Attach, you can attach the volume to up to 16 Instances built on the Nitro System in the same Availability Zone. This parameter is supported with io1 and io2 volumes only. For more information, see Amazon EBS Multi-Attach in the Amazon Elastic Compute Cloud User Guide.

createVolume_clientToken :: Lens' CreateVolume (Maybe Text)

Unique, case-sensitive identifier that you provide to ensure the idempotency of the request. For more information, see Ensure Idempotency.

createVolume_throughput :: Lens' CreateVolume (Maybe Int)

The throughput to provision for a volume, with a maximum of 1,000 MiB/s.

This parameter is valid only for gp3 volumes.

Valid Range: Minimum value of 125. Maximum value of 1000.

createVolume_size :: Lens' CreateVolume (Maybe Int)

The size of the volume, in GiBs. You must specify either a snapshot ID or a volume size. If you specify a snapshot, the default is the snapshot size. You can specify a volume size that is equal to or larger than the snapshot size.

The following are the supported volumes sizes for each volume type:

gp2 and gp3: 1-16,384

io1 and io2: 4-16,384

st1 and sc1: 125-16,384

standard: 1-1,024

createVolume_iops :: Lens' CreateVolume (Maybe Int)

The number of I/O operations per second (IOPS). For gp3, io1, and io2 volumes, this represents the number of IOPS that are provisioned for the volume. For gp2 volumes, this represents the baseline performance of the volume and the rate at which the volume accumulates I/O credits for bursting.

The following are the supported values for each volume type:

gp3: 3,000-16,000 IOPS

io1: 100-64,000 IOPS

io2: 100-64,000 IOPS

io1 and io2 volumes support up to 64,000 IOPS only on Instances built on the Nitro System. Other instance families support performance up to 32,000 IOPS.

This parameter is required for io1 and io2 volumes. The default for gp3 volumes is 3,000 IOPS. This parameter is not supported for gp2, st1, sc1, or standard volumes.

createVolume_outpostArn :: Lens' CreateVolume (Maybe Text)

The Amazon Resource Name (ARN) of the Outpost.

createVolume_encrypted :: Lens' CreateVolume (Maybe Bool)

Indicates whether the volume should be encrypted. The effect of setting the encryption state to true depends on the volume origin (new or from a snapshot), starting encryption state, ownership, and whether encryption by default is enabled. For more information, see Encryption by default in the Amazon Elastic Compute Cloud User Guide.

Encrypted Amazon EBS volumes must be attached to instances that support Amazon EBS encryption. For more information, see Supported instance types.

createVolume_tagSpecifications :: Lens' CreateVolume (Maybe [TagSpecification])

The tags to apply to the volume during creation.

createVolume_kmsKeyId :: Lens' CreateVolume (Maybe Text)

The identifier of the Key Management Service (KMS) KMS key to use for Amazon EBS encryption. If this parameter is not specified, your KMS key for Amazon EBS is used. If KmsKeyId is specified, the encrypted state must be true.

You can specify the KMS key using any of the following:

Key ID. For example, 1234abcd-12ab-34cd-56ef-1234567890ab.

Key alias. For example, alias/ExampleAlias.

Key ARN. For example, arn:aws:kms:us-east-1:012345678910:key/1234abcd-12ab-34cd-56ef-1234567890ab.

Alias ARN. For example, arn:aws:kms:us-east-1:012345678910:alias/ExampleAlias.

Amazon Web Services authenticates the KMS key asynchronously. Therefore, if you specify an ID, alias, or ARN that is not valid, the action can appear to complete, but eventually fails.

createVolume_volumeType :: Lens' CreateVolume (Maybe VolumeType)

The volume type. This parameter can be one of the following values:

General Purpose SSD: gp2 | gp3

Provisioned IOPS SSD: io1 | io2

Throughput Optimized HDD: st1

Cold HDD: sc1

Magnetic: standard

For more information, see Amazon EBS volume types in the Amazon Elastic Compute Cloud User Guide.

Default: gp2

createVolume_dryRun :: Lens' CreateVolume (Maybe Bool)

Checks whether you have the required permissions for the action, without actually making the request, and provides an error response. If you have the required permissions, the error response is DryRunOperation. Otherwise, it is UnauthorizedOperation.

createVolume_snapshotId :: Lens' CreateVolume (Maybe Text)

The snapshot from which to create the volume. You must specify either a snapshot ID or a volume size.

createVolume_availabilityZone :: Lens' CreateVolume Text

The Availability Zone in which to create the volume.

Destructuring the Response

data Volume

Describes a volume.

See: newVolume smart constructor.

Constructors

Volume'

Fields

fastRestored :: Maybe Bool

Indicates whether the volume was created using fast snapshot restore.

multiAttachEnabled :: Maybe Bool

Indicates whether Amazon EBS Multi-Attach is enabled.

attachments :: Maybe [VolumeAttachment]

Information about the volume attachments.

throughput :: Maybe Int

The throughput that the volume supports, in MiB/s.

iops :: Maybe Int

The number of I/O operations per second (IOPS). For gp3, io1, and io2 volumes, this represents the number of IOPS that are provisioned for the volume. For gp2 volumes, this represents the baseline performance of the volume and the rate at which the volume accumulates I/O credits for bursting.

outpostArn :: Maybe Text

The Amazon Resource Name (ARN) of the Outpost.

kmsKeyId :: Maybe Text

The Amazon Resource Name (ARN) of the Key Management Service (KMS) KMS key that was used to protect the volume encryption key for the volume.

tags :: Maybe [Tag]

Any tags assigned to the volume.

availabilityZone :: Text

The Availability Zone for the volume.

createTime :: ISO8601

The time stamp when volume creation was initiated.

encrypted :: Bool

Indicates whether the volume is encrypted.

size :: Int

The size of the volume, in GiBs.

snapshotId :: Text

The snapshot from which the volume was created, if applicable.

state :: VolumeState

The volume state.

volumeId :: Text

The ID of the volume.

volumeType :: VolumeType

The volume type.

newVolume

Arguments

:: Text
$sel:availabilityZone:Volume'

-> UTCTime
$sel:createTime:Volume'

-> Bool
$sel:encrypted:Volume'

-> Int
$sel:size:Volume'

-> Text
$sel:snapshotId:Volume'

-> VolumeState
$sel:state:Volume'

-> Text
$sel:volumeId:Volume'

-> VolumeType
$sel:volumeType:Volume'

-> Volume

Create a value of Volume with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:fastRestored:Volume', volume_fastRestored - Indicates whether the volume was created using fast snapshot restore.

$sel:multiAttachEnabled:Volume', volume_multiAttachEnabled - Indicates whether Amazon EBS Multi-Attach is enabled.

$sel:attachments:Volume', volume_attachments - Information about the volume attachments.

$sel:throughput:Volume', volume_throughput - The throughput that the volume supports, in MiB/s.

$sel:iops:Volume', volume_iops - The number of I/O operations per second (IOPS). For gp3, io1, and io2 volumes, this represents the number of IOPS that are provisioned for the volume. For gp2 volumes, this represents the baseline performance of the volume and the rate at which the volume accumulates I/O credits for bursting.

$sel:outpostArn:Volume', volume_outpostArn - The Amazon Resource Name (ARN) of the Outpost.

$sel:kmsKeyId:Volume', volume_kmsKeyId - The Amazon Resource Name (ARN) of the Key Management Service (KMS) KMS key that was used to protect the volume encryption key for the volume.

$sel:tags:Volume', volume_tags - Any tags assigned to the volume.

$sel:availabilityZone:Volume', volume_availabilityZone - The Availability Zone for the volume.

$sel:createTime:Volume', volume_createTime - The time stamp when volume creation was initiated.

$sel:encrypted:Volume', volume_encrypted - Indicates whether the volume is encrypted.

$sel:size:Volume', volume_size - The size of the volume, in GiBs.

$sel:snapshotId:Volume', volume_snapshotId - The snapshot from which the volume was created, if applicable.

$sel:state:Volume', volume_state - The volume state.

$sel:volumeId:Volume', volume_volumeId - The ID of the volume.

$sel:volumeType:Volume', volume_volumeType - The volume type.

Response Lenses

volume_fastRestored :: Lens' Volume (Maybe Bool)

Indicates whether the volume was created using fast snapshot restore.

volume_multiAttachEnabled :: Lens' Volume (Maybe Bool)

Indicates whether Amazon EBS Multi-Attach is enabled.

volume_attachments :: Lens' Volume (Maybe [VolumeAttachment])

Information about the volume attachments.

volume_throughput :: Lens' Volume (Maybe Int)

The throughput that the volume supports, in MiB/s.

volume_iops :: Lens' Volume (Maybe Int)

The number of I/O operations per second (IOPS). For gp3, io1, and io2 volumes, this represents the number of IOPS that are provisioned for the volume. For gp2 volumes, this represents the baseline performance of the volume and the rate at which the volume accumulates I/O credits for bursting.

volume_outpostArn :: Lens' Volume (Maybe Text)

The Amazon Resource Name (ARN) of the Outpost.

volume_kmsKeyId :: Lens' Volume (Maybe Text)

The Amazon Resource Name (ARN) of the Key Management Service (KMS) KMS key that was used to protect the volume encryption key for the volume.

volume_tags :: Lens' Volume (Maybe [Tag])

Any tags assigned to the volume.

volume_availabilityZone :: Lens' Volume Text

The Availability Zone for the volume.

volume_createTime :: Lens' Volume UTCTime

The time stamp when volume creation was initiated.

volume_encrypted :: Lens' Volume Bool

Indicates whether the volume is encrypted.

volume_size :: Lens' Volume Int

The size of the volume, in GiBs.

volume_snapshotId :: Lens' Volume Text

The snapshot from which the volume was created, if applicable.

volume_state :: Lens' Volume VolumeState

The volume state.

volume_volumeId :: Lens' Volume Text

The ID of the volume.

volume_volumeType :: Lens' Volume VolumeType

The volume type.