Amazonka.EC2.CreateVpcEndpointServiceConfiguration

Description

Creates a VPC endpoint service configuration to which service consumers (Amazon Web Services accounts, IAM users, and IAM roles) can connect.

To create an endpoint service configuration, you must first create one of the following for your service:

A Network Load Balancer. Service consumers connect to your service using an interface endpoint.

A Gateway Load Balancer. Service consumers connect to your service using a Gateway Load Balancer endpoint.

For more information, see VPC Endpoint Services in the Amazon Virtual Private Cloud User Guide.

If you set the private DNS name, you must prove that you own the private DNS domain name. For more information, see VPC Endpoint Service Private DNS Name Verification in the Amazon Virtual Private Cloud User Guide.

Creating a Request

data CreateVpcEndpointServiceConfiguration

See: newCreateVpcEndpointServiceConfiguration smart constructor.

Constructors

CreateVpcEndpointServiceConfiguration'

Fields

networkLoadBalancerArns :: Maybe [Text]

The Amazon Resource Names (ARNs) of one or more Network Load Balancers for your service.

clientToken :: Maybe Text

Unique, case-sensitive identifier that you provide to ensure the idempotency of the request. For more information, see How to ensure idempotency.

tagSpecifications :: Maybe [TagSpecification]

The tags to associate with the service.

gatewayLoadBalancerArns :: Maybe [Text]

The Amazon Resource Names (ARNs) of one or more Gateway Load Balancers.

acceptanceRequired :: Maybe Bool

Indicates whether requests from service consumers to create an endpoint to your service must be accepted. To accept a request, use AcceptVpcEndpointConnections.

privateDnsName :: Maybe Text

(Interface endpoint configuration) The private DNS name to assign to the VPC endpoint service.

dryRun :: Maybe Bool

Checks whether you have the required permissions for the action, without actually making the request, and provides an error response. If you have the required permissions, the error response is DryRunOperation. Otherwise, it is UnauthorizedOperation.

newCreateVpcEndpointServiceConfiguration :: CreateVpcEndpointServiceConfiguration

Create a value of CreateVpcEndpointServiceConfiguration with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:networkLoadBalancerArns:CreateVpcEndpointServiceConfiguration', createVpcEndpointServiceConfiguration_networkLoadBalancerArns - The Amazon Resource Names (ARNs) of one or more Network Load Balancers for your service.

$sel:clientToken:CreateVpcEndpointServiceConfiguration', createVpcEndpointServiceConfiguration_clientToken - Unique, case-sensitive identifier that you provide to ensure the idempotency of the request. For more information, see How to ensure idempotency.

$sel:tagSpecifications:CreateVpcEndpointServiceConfiguration', createVpcEndpointServiceConfiguration_tagSpecifications - The tags to associate with the service.

$sel:gatewayLoadBalancerArns:CreateVpcEndpointServiceConfiguration', createVpcEndpointServiceConfiguration_gatewayLoadBalancerArns - The Amazon Resource Names (ARNs) of one or more Gateway Load Balancers.

$sel:acceptanceRequired:CreateVpcEndpointServiceConfiguration', createVpcEndpointServiceConfiguration_acceptanceRequired - Indicates whether requests from service consumers to create an endpoint to your service must be accepted. To accept a request, use AcceptVpcEndpointConnections.

$sel:privateDnsName:CreateVpcEndpointServiceConfiguration', createVpcEndpointServiceConfiguration_privateDnsName - (Interface endpoint configuration) The private DNS name to assign to the VPC endpoint service.

$sel:dryRun:CreateVpcEndpointServiceConfiguration', createVpcEndpointServiceConfiguration_dryRun - Checks whether you have the required permissions for the action, without actually making the request, and provides an error response. If you have the required permissions, the error response is DryRunOperation. Otherwise, it is UnauthorizedOperation.

Request Lenses

createVpcEndpointServiceConfiguration_networkLoadBalancerArns :: Lens' CreateVpcEndpointServiceConfiguration (Maybe [Text])

The Amazon Resource Names (ARNs) of one or more Network Load Balancers for your service.

createVpcEndpointServiceConfiguration_clientToken :: Lens' CreateVpcEndpointServiceConfiguration (Maybe Text)

Unique, case-sensitive identifier that you provide to ensure the idempotency of the request. For more information, see How to ensure idempotency.

createVpcEndpointServiceConfiguration_tagSpecifications :: Lens' CreateVpcEndpointServiceConfiguration (Maybe [TagSpecification])

The tags to associate with the service.

createVpcEndpointServiceConfiguration_gatewayLoadBalancerArns :: Lens' CreateVpcEndpointServiceConfiguration (Maybe [Text])

The Amazon Resource Names (ARNs) of one or more Gateway Load Balancers.

createVpcEndpointServiceConfiguration_acceptanceRequired :: Lens' CreateVpcEndpointServiceConfiguration (Maybe Bool)

Indicates whether requests from service consumers to create an endpoint to your service must be accepted. To accept a request, use AcceptVpcEndpointConnections.

createVpcEndpointServiceConfiguration_privateDnsName :: Lens' CreateVpcEndpointServiceConfiguration (Maybe Text)

(Interface endpoint configuration) The private DNS name to assign to the VPC endpoint service.

createVpcEndpointServiceConfiguration_dryRun :: Lens' CreateVpcEndpointServiceConfiguration (Maybe Bool)

Checks whether you have the required permissions for the action, without actually making the request, and provides an error response. If you have the required permissions, the error response is DryRunOperation. Otherwise, it is UnauthorizedOperation.

Destructuring the Response

data CreateVpcEndpointServiceConfigurationResponse

See: newCreateVpcEndpointServiceConfigurationResponse smart constructor.

Constructors

CreateVpcEndpointServiceConfigurationResponse'

Fields

clientToken :: Maybe Text

Unique, case-sensitive identifier that you provide to ensure the idempotency of the request.

serviceConfiguration :: Maybe ServiceConfiguration

Information about the service configuration.

httpStatus :: Int

The response's http status code.

newCreateVpcEndpointServiceConfigurationResponse

Arguments

:: Int
$sel:httpStatus:CreateVpcEndpointServiceConfigurationResponse'

-> CreateVpcEndpointServiceConfigurationResponse

Create a value of CreateVpcEndpointServiceConfigurationResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:clientToken:CreateVpcEndpointServiceConfiguration', createVpcEndpointServiceConfigurationResponse_clientToken - Unique, case-sensitive identifier that you provide to ensure the idempotency of the request.

$sel:serviceConfiguration:CreateVpcEndpointServiceConfigurationResponse', createVpcEndpointServiceConfigurationResponse_serviceConfiguration - Information about the service configuration.

$sel:httpStatus:CreateVpcEndpointServiceConfigurationResponse', createVpcEndpointServiceConfigurationResponse_httpStatus - The response's http status code.

Response Lenses

createVpcEndpointServiceConfigurationResponse_clientToken :: Lens' CreateVpcEndpointServiceConfigurationResponse (Maybe Text)

Unique, case-sensitive identifier that you provide to ensure the idempotency of the request.

createVpcEndpointServiceConfigurationResponse_serviceConfiguration :: Lens' CreateVpcEndpointServiceConfigurationResponse (Maybe ServiceConfiguration)

Information about the service configuration.

createVpcEndpointServiceConfigurationResponse_httpStatus :: Lens' CreateVpcEndpointServiceConfigurationResponse Int

The response's http status code.