Amazonka.EC2.CreateVpnGateway

Description

Creates a virtual private gateway. A virtual private gateway is the endpoint on the VPC side of your VPN connection. You can create a virtual private gateway before creating the VPC itself.

For more information, see Amazon Web Services Site-to-Site VPN in the Amazon Web Services Site-to-Site VPN User Guide.

Creating a Request

data CreateVpnGateway

Contains the parameters for CreateVpnGateway.

See: newCreateVpnGateway smart constructor.

Constructors

CreateVpnGateway'

Fields

amazonSideAsn :: Maybe Integer

A private Autonomous System Number (ASN) for the Amazon side of a BGP session. If you're using a 16-bit ASN, it must be in the 64512 to 65534 range. If you're using a 32-bit ASN, it must be in the 4200000000 to 4294967294 range.

Default: 64512

tagSpecifications :: Maybe [TagSpecification]

The tags to apply to the virtual private gateway.

availabilityZone :: Maybe Text

The Availability Zone for the virtual private gateway.

dryRun :: Maybe Bool

Checks whether you have the required permissions for the action, without actually making the request, and provides an error response. If you have the required permissions, the error response is DryRunOperation. Otherwise, it is UnauthorizedOperation.

type' :: GatewayType

The type of VPN connection this virtual private gateway supports.

newCreateVpnGateway

Arguments

:: GatewayType
$sel:type':CreateVpnGateway'

-> CreateVpnGateway

Create a value of CreateVpnGateway with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:amazonSideAsn:CreateVpnGateway', createVpnGateway_amazonSideAsn - A private Autonomous System Number (ASN) for the Amazon side of a BGP session. If you're using a 16-bit ASN, it must be in the 64512 to 65534 range. If you're using a 32-bit ASN, it must be in the 4200000000 to 4294967294 range.

Default: 64512

$sel:tagSpecifications:CreateVpnGateway', createVpnGateway_tagSpecifications - The tags to apply to the virtual private gateway.

$sel:availabilityZone:CreateVpnGateway', createVpnGateway_availabilityZone - The Availability Zone for the virtual private gateway.

$sel:dryRun:CreateVpnGateway', createVpnGateway_dryRun - Checks whether you have the required permissions for the action, without actually making the request, and provides an error response. If you have the required permissions, the error response is DryRunOperation. Otherwise, it is UnauthorizedOperation.

$sel:type':CreateVpnGateway', createVpnGateway_type - The type of VPN connection this virtual private gateway supports.

Request Lenses

createVpnGateway_amazonSideAsn :: Lens' CreateVpnGateway (Maybe Integer)

A private Autonomous System Number (ASN) for the Amazon side of a BGP session. If you're using a 16-bit ASN, it must be in the 64512 to 65534 range. If you're using a 32-bit ASN, it must be in the 4200000000 to 4294967294 range.

Default: 64512

createVpnGateway_tagSpecifications :: Lens' CreateVpnGateway (Maybe [TagSpecification])

The tags to apply to the virtual private gateway.

createVpnGateway_availabilityZone :: Lens' CreateVpnGateway (Maybe Text)

The Availability Zone for the virtual private gateway.

createVpnGateway_dryRun :: Lens' CreateVpnGateway (Maybe Bool)

Checks whether you have the required permissions for the action, without actually making the request, and provides an error response. If you have the required permissions, the error response is DryRunOperation. Otherwise, it is UnauthorizedOperation.

createVpnGateway_type :: Lens' CreateVpnGateway GatewayType

The type of VPN connection this virtual private gateway supports.

Destructuring the Response

data CreateVpnGatewayResponse

Contains the output of CreateVpnGateway.

See: newCreateVpnGatewayResponse smart constructor.

Constructors

CreateVpnGatewayResponse'

Fields

vpnGateway :: Maybe VpnGateway

Information about the virtual private gateway.

httpStatus :: Int

The response's http status code.

newCreateVpnGatewayResponse

Arguments

:: Int
$sel:httpStatus:CreateVpnGatewayResponse'

-> CreateVpnGatewayResponse

Create a value of CreateVpnGatewayResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:vpnGateway:CreateVpnGatewayResponse', createVpnGatewayResponse_vpnGateway - Information about the virtual private gateway.

$sel:httpStatus:CreateVpnGatewayResponse', createVpnGatewayResponse_httpStatus - The response's http status code.

Response Lenses

createVpnGatewayResponse_vpnGateway :: Lens' CreateVpnGatewayResponse (Maybe VpnGateway)

Information about the virtual private gateway.

createVpnGatewayResponse_httpStatus :: Lens' CreateVpnGatewayResponse Int

The response's http status code.