Amazonka.EC2.DeleteFleets

Description

Deletes the specified EC2 Fleet.

After you delete an EC2 Fleet, it launches no new instances.

You must specify whether a deleted EC2 Fleet should also terminate its instances. If you choose to terminate the instances, the EC2 Fleet enters the deleted_terminating state. Otherwise, the EC2 Fleet enters the deleted_running state, and the instances continue to run until they are interrupted or you terminate them manually.

For instant fleets, EC2 Fleet must terminate the instances when the fleet is deleted. A deleted instant fleet with running instances is not supported.

Restrictions

You can delete up to 25 instant fleets in a single request. If you exceed this number, no instant fleets are deleted and an error is returned. There is no restriction on the number of fleets of type maintain or request that can be deleted in a single request.

Up to 1000 instances can be terminated in a single request to delete instant fleets.

For more information, see Deleting an EC2 Fleet in the Amazon EC2 User Guide.

Creating a Request

data DeleteFleets

See: newDeleteFleets smart constructor.

Constructors

DeleteFleets'

Fields

dryRun :: Maybe Bool

Checks whether you have the required permissions for the action, without actually making the request, and provides an error response. If you have the required permissions, the error response is DryRunOperation. Otherwise, it is UnauthorizedOperation.

fleetIds :: [Text]

The IDs of the EC2 Fleets.

terminateInstances :: Bool

Indicates whether to terminate the instances when the EC2 Fleet is deleted. The default is to terminate the instances.

To let the instances continue to run after the EC2 Fleet is deleted, specify NoTerminateInstances. Supported only for fleets of type maintain and request.

For instant fleets, you cannot specify NoTerminateInstances. A deleted instant fleet with running instances is not supported.

newDeleteFleets

Arguments

:: Bool
$sel:terminateInstances:DeleteFleets'

-> DeleteFleets

Create a value of DeleteFleets with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:dryRun:DeleteFleets', deleteFleets_dryRun - Checks whether you have the required permissions for the action, without actually making the request, and provides an error response. If you have the required permissions, the error response is DryRunOperation. Otherwise, it is UnauthorizedOperation.

$sel:fleetIds:DeleteFleets', deleteFleets_fleetIds - The IDs of the EC2 Fleets.

$sel:terminateInstances:DeleteFleets', deleteFleets_terminateInstances - Indicates whether to terminate the instances when the EC2 Fleet is deleted. The default is to terminate the instances.

To let the instances continue to run after the EC2 Fleet is deleted, specify NoTerminateInstances. Supported only for fleets of type maintain and request.

For instant fleets, you cannot specify NoTerminateInstances. A deleted instant fleet with running instances is not supported.

Request Lenses

deleteFleets_dryRun :: Lens' DeleteFleets (Maybe Bool)

Checks whether you have the required permissions for the action, without actually making the request, and provides an error response. If you have the required permissions, the error response is DryRunOperation. Otherwise, it is UnauthorizedOperation.

deleteFleets_fleetIds :: Lens' DeleteFleets [Text]

The IDs of the EC2 Fleets.

deleteFleets_terminateInstances :: Lens' DeleteFleets Bool

Indicates whether to terminate the instances when the EC2 Fleet is deleted. The default is to terminate the instances.

To let the instances continue to run after the EC2 Fleet is deleted, specify NoTerminateInstances. Supported only for fleets of type maintain and request.

For instant fleets, you cannot specify NoTerminateInstances. A deleted instant fleet with running instances is not supported.

Destructuring the Response

data DeleteFleetsResponse

See: newDeleteFleetsResponse smart constructor.

Constructors

DeleteFleetsResponse'

Fields

successfulFleetDeletions :: Maybe [DeleteFleetSuccessItem]

Information about the EC2 Fleets that are successfully deleted.

unsuccessfulFleetDeletions :: Maybe [DeleteFleetErrorItem]

Information about the EC2 Fleets that are not successfully deleted.

httpStatus :: Int

The response's http status code.

newDeleteFleetsResponse

Arguments

:: Int
$sel:httpStatus:DeleteFleetsResponse'

-> DeleteFleetsResponse

Create a value of DeleteFleetsResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:successfulFleetDeletions:DeleteFleetsResponse', deleteFleetsResponse_successfulFleetDeletions - Information about the EC2 Fleets that are successfully deleted.

$sel:unsuccessfulFleetDeletions:DeleteFleetsResponse', deleteFleetsResponse_unsuccessfulFleetDeletions - Information about the EC2 Fleets that are not successfully deleted.

$sel:httpStatus:DeleteFleetsResponse', deleteFleetsResponse_httpStatus - The response's http status code.

Response Lenses

deleteFleetsResponse_successfulFleetDeletions :: Lens' DeleteFleetsResponse (Maybe [DeleteFleetSuccessItem])

Information about the EC2 Fleets that are successfully deleted.

deleteFleetsResponse_unsuccessfulFleetDeletions :: Lens' DeleteFleetsResponse (Maybe [DeleteFleetErrorItem])

Information about the EC2 Fleets that are not successfully deleted.

deleteFleetsResponse_httpStatus :: Lens' DeleteFleetsResponse Int

The response's http status code.