Amazonka.EC2.DeleteNatGateway

Description

Deletes the specified NAT gateway. Deleting a public NAT gateway disassociates its Elastic IP address, but does not release the address from your account. Deleting a NAT gateway does not delete any NAT gateway routes in your route tables.

Creating a Request

data DeleteNatGateway

See: newDeleteNatGateway smart constructor.

Constructors

DeleteNatGateway'

Fields

dryRun :: Maybe Bool

Checks whether you have the required permissions for the action, without actually making the request, and provides an error response. If you have the required permissions, the error response is DryRunOperation. Otherwise, it is UnauthorizedOperation.

natGatewayId :: Text

The ID of the NAT gateway.

newDeleteNatGateway

Arguments

:: Text
$sel:natGatewayId:DeleteNatGateway'

-> DeleteNatGateway

Create a value of DeleteNatGateway with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:dryRun:DeleteNatGateway', deleteNatGateway_dryRun - Checks whether you have the required permissions for the action, without actually making the request, and provides an error response. If you have the required permissions, the error response is DryRunOperation. Otherwise, it is UnauthorizedOperation.

$sel:natGatewayId:DeleteNatGateway', deleteNatGateway_natGatewayId - The ID of the NAT gateway.

Request Lenses

deleteNatGateway_dryRun :: Lens' DeleteNatGateway (Maybe Bool)

Checks whether you have the required permissions for the action, without actually making the request, and provides an error response. If you have the required permissions, the error response is DryRunOperation. Otherwise, it is UnauthorizedOperation.

deleteNatGateway_natGatewayId :: Lens' DeleteNatGateway Text

The ID of the NAT gateway.

Destructuring the Response

data DeleteNatGatewayResponse

See: newDeleteNatGatewayResponse smart constructor.

Constructors

DeleteNatGatewayResponse'

Fields

natGatewayId :: Maybe Text

The ID of the NAT gateway.

httpStatus :: Int

The response's http status code.

newDeleteNatGatewayResponse

Arguments

:: Int
$sel:httpStatus:DeleteNatGatewayResponse'

-> DeleteNatGatewayResponse

Create a value of DeleteNatGatewayResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:natGatewayId:DeleteNatGateway', deleteNatGatewayResponse_natGatewayId - The ID of the NAT gateway.

$sel:httpStatus:DeleteNatGatewayResponse', deleteNatGatewayResponse_httpStatus - The response's http status code.

Response Lenses

deleteNatGatewayResponse_natGatewayId :: Lens' DeleteNatGatewayResponse (Maybe Text)

The ID of the NAT gateway.

deleteNatGatewayResponse_httpStatus :: Lens' DeleteNatGatewayResponse Int

The response's http status code.