Amazonka.EC2.DeleteNetworkInsightsAnalysis

Description

Deletes the specified network insights analysis.

Creating a Request

data DeleteNetworkInsightsAnalysis

See: newDeleteNetworkInsightsAnalysis smart constructor.

Constructors

DeleteNetworkInsightsAnalysis'

Fields

dryRun :: Maybe Bool

Checks whether you have the required permissions for the action, without actually making the request, and provides an error response. If you have the required permissions, the error response is DryRunOperation. Otherwise, it is UnauthorizedOperation.

networkInsightsAnalysisId :: Text

The ID of the network insights analysis.

newDeleteNetworkInsightsAnalysis

Arguments

:: Text
$sel:networkInsightsAnalysisId:DeleteNetworkInsightsAnalysis'

-> DeleteNetworkInsightsAnalysis

Create a value of DeleteNetworkInsightsAnalysis with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:dryRun:DeleteNetworkInsightsAnalysis', deleteNetworkInsightsAnalysis_dryRun - Checks whether you have the required permissions for the action, without actually making the request, and provides an error response. If you have the required permissions, the error response is DryRunOperation. Otherwise, it is UnauthorizedOperation.

$sel:networkInsightsAnalysisId:DeleteNetworkInsightsAnalysis', deleteNetworkInsightsAnalysis_networkInsightsAnalysisId - The ID of the network insights analysis.

Request Lenses

deleteNetworkInsightsAnalysis_dryRun :: Lens' DeleteNetworkInsightsAnalysis (Maybe Bool)

Checks whether you have the required permissions for the action, without actually making the request, and provides an error response. If you have the required permissions, the error response is DryRunOperation. Otherwise, it is UnauthorizedOperation.

deleteNetworkInsightsAnalysis_networkInsightsAnalysisId :: Lens' DeleteNetworkInsightsAnalysis Text

The ID of the network insights analysis.

Destructuring the Response

data DeleteNetworkInsightsAnalysisResponse

See: newDeleteNetworkInsightsAnalysisResponse smart constructor.

Constructors

DeleteNetworkInsightsAnalysisResponse'

Fields

networkInsightsAnalysisId :: Maybe Text

The ID of the network insights analysis.

httpStatus :: Int

The response's http status code.

newDeleteNetworkInsightsAnalysisResponse

Arguments

:: Int
$sel:httpStatus:DeleteNetworkInsightsAnalysisResponse'

-> DeleteNetworkInsightsAnalysisResponse

Create a value of DeleteNetworkInsightsAnalysisResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:networkInsightsAnalysisId:DeleteNetworkInsightsAnalysis', deleteNetworkInsightsAnalysisResponse_networkInsightsAnalysisId - The ID of the network insights analysis.

$sel:httpStatus:DeleteNetworkInsightsAnalysisResponse', deleteNetworkInsightsAnalysisResponse_httpStatus - The response's http status code.

Response Lenses

deleteNetworkInsightsAnalysisResponse_networkInsightsAnalysisId :: Lens' DeleteNetworkInsightsAnalysisResponse (Maybe Text)

The ID of the network insights analysis.

deleteNetworkInsightsAnalysisResponse_httpStatus :: Lens' DeleteNetworkInsightsAnalysisResponse Int

The response's http status code.