Amazonka.EC2.DeleteTransitGateway

Description

Deletes the specified transit gateway.

Creating a Request

data DeleteTransitGateway

See: newDeleteTransitGateway smart constructor.

Constructors

DeleteTransitGateway'

Fields

dryRun :: Maybe Bool

Checks whether you have the required permissions for the action, without actually making the request, and provides an error response. If you have the required permissions, the error response is DryRunOperation. Otherwise, it is UnauthorizedOperation.

transitGatewayId :: Text

The ID of the transit gateway.

newDeleteTransitGateway

Arguments

:: Text
$sel:transitGatewayId:DeleteTransitGateway'

-> DeleteTransitGateway

Create a value of DeleteTransitGateway with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:dryRun:DeleteTransitGateway', deleteTransitGateway_dryRun - Checks whether you have the required permissions for the action, without actually making the request, and provides an error response. If you have the required permissions, the error response is DryRunOperation. Otherwise, it is UnauthorizedOperation.

$sel:transitGatewayId:DeleteTransitGateway', deleteTransitGateway_transitGatewayId - The ID of the transit gateway.

Request Lenses

deleteTransitGateway_dryRun :: Lens' DeleteTransitGateway (Maybe Bool)

Checks whether you have the required permissions for the action, without actually making the request, and provides an error response. If you have the required permissions, the error response is DryRunOperation. Otherwise, it is UnauthorizedOperation.

deleteTransitGateway_transitGatewayId :: Lens' DeleteTransitGateway Text

The ID of the transit gateway.

Destructuring the Response

data DeleteTransitGatewayResponse

See: newDeleteTransitGatewayResponse smart constructor.

Constructors

DeleteTransitGatewayResponse'

Fields

transitGateway :: Maybe TransitGateway

Information about the deleted transit gateway.

httpStatus :: Int

The response's http status code.

newDeleteTransitGatewayResponse

Arguments

:: Int
$sel:httpStatus:DeleteTransitGatewayResponse'

-> DeleteTransitGatewayResponse

Create a value of DeleteTransitGatewayResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:transitGateway:DeleteTransitGatewayResponse', deleteTransitGatewayResponse_transitGateway - Information about the deleted transit gateway.

$sel:httpStatus:DeleteTransitGatewayResponse', deleteTransitGatewayResponse_httpStatus - The response's http status code.

Response Lenses

deleteTransitGatewayResponse_transitGateway :: Lens' DeleteTransitGatewayResponse (Maybe TransitGateway)

Information about the deleted transit gateway.

deleteTransitGatewayResponse_httpStatus :: Lens' DeleteTransitGatewayResponse Int

The response's http status code.