Amazonka.EC2.DeleteTransitGatewayRoute

Description

Deletes the specified route from the specified transit gateway route table.

Creating a Request

data DeleteTransitGatewayRoute

See: newDeleteTransitGatewayRoute smart constructor.

Constructors

DeleteTransitGatewayRoute'

Fields

dryRun :: Maybe Bool

Checks whether you have the required permissions for the action, without actually making the request, and provides an error response. If you have the required permissions, the error response is DryRunOperation. Otherwise, it is UnauthorizedOperation.

transitGatewayRouteTableId :: Text

The ID of the transit gateway route table.

destinationCidrBlock :: Text

The CIDR range for the route. This must match the CIDR for the route exactly.

newDeleteTransitGatewayRoute

Arguments

:: Text
$sel:transitGatewayRouteTableId:DeleteTransitGatewayRoute'

-> Text
$sel:destinationCidrBlock:DeleteTransitGatewayRoute'

-> DeleteTransitGatewayRoute

Create a value of DeleteTransitGatewayRoute with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:dryRun:DeleteTransitGatewayRoute', deleteTransitGatewayRoute_dryRun - Checks whether you have the required permissions for the action, without actually making the request, and provides an error response. If you have the required permissions, the error response is DryRunOperation. Otherwise, it is UnauthorizedOperation.

$sel:transitGatewayRouteTableId:DeleteTransitGatewayRoute', deleteTransitGatewayRoute_transitGatewayRouteTableId - The ID of the transit gateway route table.

$sel:destinationCidrBlock:DeleteTransitGatewayRoute', deleteTransitGatewayRoute_destinationCidrBlock - The CIDR range for the route. This must match the CIDR for the route exactly.

Request Lenses

deleteTransitGatewayRoute_dryRun :: Lens' DeleteTransitGatewayRoute (Maybe Bool)

Checks whether you have the required permissions for the action, without actually making the request, and provides an error response. If you have the required permissions, the error response is DryRunOperation. Otherwise, it is UnauthorizedOperation.

deleteTransitGatewayRoute_transitGatewayRouteTableId :: Lens' DeleteTransitGatewayRoute Text

The ID of the transit gateway route table.

deleteTransitGatewayRoute_destinationCidrBlock :: Lens' DeleteTransitGatewayRoute Text

The CIDR range for the route. This must match the CIDR for the route exactly.

Destructuring the Response

data DeleteTransitGatewayRouteResponse

See: newDeleteTransitGatewayRouteResponse smart constructor.

Constructors

DeleteTransitGatewayRouteResponse'

Fields

route :: Maybe TransitGatewayRoute

Information about the route.

httpStatus :: Int

The response's http status code.

newDeleteTransitGatewayRouteResponse

Arguments

:: Int
$sel:httpStatus:DeleteTransitGatewayRouteResponse'

-> DeleteTransitGatewayRouteResponse

Create a value of DeleteTransitGatewayRouteResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:route:DeleteTransitGatewayRouteResponse', deleteTransitGatewayRouteResponse_route - Information about the route.

$sel:httpStatus:DeleteTransitGatewayRouteResponse', deleteTransitGatewayRouteResponse_httpStatus - The response's http status code.

Response Lenses

deleteTransitGatewayRouteResponse_route :: Lens' DeleteTransitGatewayRouteResponse (Maybe TransitGatewayRoute)

Information about the route.

deleteTransitGatewayRouteResponse_httpStatus :: Lens' DeleteTransitGatewayRouteResponse Int

The response's http status code.