Amazonka.EC2.DeleteVpc

Description

Deletes the specified VPC. You must detach or delete all gateways and resources that are associated with the VPC before you can delete it. For example, you must terminate all instances running in the VPC, delete all security groups associated with the VPC (except the default one), delete all route tables associated with the VPC (except the default one), and so on.

Creating a Request

data DeleteVpc

See: newDeleteVpc smart constructor.

Constructors

DeleteVpc'

Fields

dryRun :: Maybe Bool

Checks whether you have the required permissions for the action, without actually making the request, and provides an error response. If you have the required permissions, the error response is DryRunOperation. Otherwise, it is UnauthorizedOperation.

vpcId :: Text

The ID of the VPC.

newDeleteVpc

Arguments

:: Text
$sel:vpcId:DeleteVpc'

-> DeleteVpc

Create a value of DeleteVpc with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:dryRun:DeleteVpc', deleteVpc_dryRun - Checks whether you have the required permissions for the action, without actually making the request, and provides an error response. If you have the required permissions, the error response is DryRunOperation. Otherwise, it is UnauthorizedOperation.

$sel:vpcId:DeleteVpc', deleteVpc_vpcId - The ID of the VPC.

Request Lenses

deleteVpc_dryRun :: Lens' DeleteVpc (Maybe Bool)

Checks whether you have the required permissions for the action, without actually making the request, and provides an error response. If you have the required permissions, the error response is DryRunOperation. Otherwise, it is UnauthorizedOperation.

deleteVpc_vpcId :: Lens' DeleteVpc Text

The ID of the VPC.

Destructuring the Response

data DeleteVpcResponse

See: newDeleteVpcResponse smart constructor.

Constructors

DeleteVpcResponse'

newDeleteVpcResponse :: DeleteVpcResponse

Create a value of DeleteVpcResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.