Amazonka.EC2.DescribeAvailabilityZones

Description

Describes the Availability Zones, Local Zones, and Wavelength Zones that are available to you. If there is an event impacting a zone, you can use this request to view the state and any provided messages for that zone.

For more information about Availability Zones, Local Zones, and Wavelength Zones, see Regions and zones in the Amazon Elastic Compute Cloud User Guide.

Creating a Request

data DescribeAvailabilityZones

See: newDescribeAvailabilityZones smart constructor.

Constructors

DescribeAvailabilityZones'

Fields

zoneNames :: Maybe [Text]

The names of the Availability Zones, Local Zones, and Wavelength Zones.

allAvailabilityZones :: Maybe Bool

Include all Availability Zones, Local Zones, and Wavelength Zones regardless of your opt-in status.

If you do not use this parameter, the results include only the zones for the Regions where you have chosen the option to opt in.

zoneIds :: Maybe [Text]

The IDs of the Availability Zones, Local Zones, and Wavelength Zones.

filters :: Maybe [Filter]

The filters.

group-name - For Availability Zones, use the Region name. For Local Zones, use the name of the group associated with the Local Zone (for example, us-west-2-lax-1) For Wavelength Zones, use the name of the group associated with the Wavelength Zone (for example, us-east-1-wl1-bos-wlz-1).

message - The Zone message.

opt-in-status - The opt-in status (opted-in, and not-opted-in | opt-in-not-required).

parent-zoneID - The ID of the zone that handles some of the Local Zone and Wavelength Zone control plane operations, such as API calls.

parent-zoneName - The ID of the zone that handles some of the Local Zone and Wavelength Zone control plane operations, such as API calls.

region-name - The name of the Region for the Zone (for example, us-east-1).

state - The state of the Availability Zone, the Local Zone, or the Wavelength Zone (available | information | impaired | unavailable).

zone-id - The ID of the Availability Zone (for example, use1-az1), the Local Zone (for example, usw2-lax1-az1), or the Wavelength Zone (for example, us-east-1-wl1-bos-wlz-1).

zone-type - The type of zone, for example, local-zone.

zone-name - The name of the Availability Zone (for example, us-east-1a), the Local Zone (for example, us-west-2-lax-1a), or the Wavelength Zone (for example, us-east-1-wl1-bos-wlz-1).

zone-type - The type of zone, for example, local-zone.

dryRun :: Maybe Bool

Checks whether you have the required permissions for the action, without actually making the request, and provides an error response. If you have the required permissions, the error response is DryRunOperation. Otherwise, it is UnauthorizedOperation.

newDescribeAvailabilityZones :: DescribeAvailabilityZones

Create a value of DescribeAvailabilityZones with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:zoneNames:DescribeAvailabilityZones', describeAvailabilityZones_zoneNames - The names of the Availability Zones, Local Zones, and Wavelength Zones.

$sel:allAvailabilityZones:DescribeAvailabilityZones', describeAvailabilityZones_allAvailabilityZones - Include all Availability Zones, Local Zones, and Wavelength Zones regardless of your opt-in status.

If you do not use this parameter, the results include only the zones for the Regions where you have chosen the option to opt in.

$sel:zoneIds:DescribeAvailabilityZones', describeAvailabilityZones_zoneIds - The IDs of the Availability Zones, Local Zones, and Wavelength Zones.

$sel:filters:DescribeAvailabilityZones', describeAvailabilityZones_filters - The filters.

group-name - For Availability Zones, use the Region name. For Local Zones, use the name of the group associated with the Local Zone (for example, us-west-2-lax-1) For Wavelength Zones, use the name of the group associated with the Wavelength Zone (for example, us-east-1-wl1-bos-wlz-1).

message - The Zone message.

opt-in-status - The opt-in status (opted-in, and not-opted-in | opt-in-not-required).

parent-zoneID - The ID of the zone that handles some of the Local Zone and Wavelength Zone control plane operations, such as API calls.

parent-zoneName - The ID of the zone that handles some of the Local Zone and Wavelength Zone control plane operations, such as API calls.

region-name - The name of the Region for the Zone (for example, us-east-1).

state - The state of the Availability Zone, the Local Zone, or the Wavelength Zone (available | information | impaired | unavailable).

zone-id - The ID of the Availability Zone (for example, use1-az1), the Local Zone (for example, usw2-lax1-az1), or the Wavelength Zone (for example, us-east-1-wl1-bos-wlz-1).

zone-type - The type of zone, for example, local-zone.

zone-name - The name of the Availability Zone (for example, us-east-1a), the Local Zone (for example, us-west-2-lax-1a), or the Wavelength Zone (for example, us-east-1-wl1-bos-wlz-1).

zone-type - The type of zone, for example, local-zone.

$sel:dryRun:DescribeAvailabilityZones', describeAvailabilityZones_dryRun - Checks whether you have the required permissions for the action, without actually making the request, and provides an error response. If you have the required permissions, the error response is DryRunOperation. Otherwise, it is UnauthorizedOperation.

Request Lenses

describeAvailabilityZones_zoneNames :: Lens' DescribeAvailabilityZones (Maybe [Text])

The names of the Availability Zones, Local Zones, and Wavelength Zones.

describeAvailabilityZones_allAvailabilityZones :: Lens' DescribeAvailabilityZones (Maybe Bool)

Include all Availability Zones, Local Zones, and Wavelength Zones regardless of your opt-in status.

If you do not use this parameter, the results include only the zones for the Regions where you have chosen the option to opt in.

describeAvailabilityZones_zoneIds :: Lens' DescribeAvailabilityZones (Maybe [Text])

The IDs of the Availability Zones, Local Zones, and Wavelength Zones.

describeAvailabilityZones_filters :: Lens' DescribeAvailabilityZones (Maybe [Filter])

The filters.

group-name - For Availability Zones, use the Region name. For Local Zones, use the name of the group associated with the Local Zone (for example, us-west-2-lax-1) For Wavelength Zones, use the name of the group associated with the Wavelength Zone (for example, us-east-1-wl1-bos-wlz-1).

message - The Zone message.

opt-in-status - The opt-in status (opted-in, and not-opted-in | opt-in-not-required).

parent-zoneID - The ID of the zone that handles some of the Local Zone and Wavelength Zone control plane operations, such as API calls.

parent-zoneName - The ID of the zone that handles some of the Local Zone and Wavelength Zone control plane operations, such as API calls.

region-name - The name of the Region for the Zone (for example, us-east-1).

state - The state of the Availability Zone, the Local Zone, or the Wavelength Zone (available | information | impaired | unavailable).

zone-id - The ID of the Availability Zone (for example, use1-az1), the Local Zone (for example, usw2-lax1-az1), or the Wavelength Zone (for example, us-east-1-wl1-bos-wlz-1).

zone-type - The type of zone, for example, local-zone.

zone-name - The name of the Availability Zone (for example, us-east-1a), the Local Zone (for example, us-west-2-lax-1a), or the Wavelength Zone (for example, us-east-1-wl1-bos-wlz-1).

zone-type - The type of zone, for example, local-zone.

describeAvailabilityZones_dryRun :: Lens' DescribeAvailabilityZones (Maybe Bool)

Checks whether you have the required permissions for the action, without actually making the request, and provides an error response. If you have the required permissions, the error response is DryRunOperation. Otherwise, it is UnauthorizedOperation.

Destructuring the Response

data DescribeAvailabilityZonesResponse

See: newDescribeAvailabilityZonesResponse smart constructor.

Constructors

DescribeAvailabilityZonesResponse'

Fields

availabilityZones :: Maybe [AvailabilityZone]

Information about the Availability Zones, Local Zones, and Wavelength Zones.

httpStatus :: Int

The response's http status code.

newDescribeAvailabilityZonesResponse

Arguments

:: Int
$sel:httpStatus:DescribeAvailabilityZonesResponse'

-> DescribeAvailabilityZonesResponse

Create a value of DescribeAvailabilityZonesResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:availabilityZones:DescribeAvailabilityZonesResponse', describeAvailabilityZonesResponse_availabilityZones - Information about the Availability Zones, Local Zones, and Wavelength Zones.

$sel:httpStatus:DescribeAvailabilityZonesResponse', describeAvailabilityZonesResponse_httpStatus - The response's http status code.

Response Lenses

describeAvailabilityZonesResponse_availabilityZones :: Lens' DescribeAvailabilityZonesResponse (Maybe [AvailabilityZone])

Information about the Availability Zones, Local Zones, and Wavelength Zones.

describeAvailabilityZonesResponse_httpStatus :: Lens' DescribeAvailabilityZonesResponse Int

The response's http status code.