Amazonka.EC2.DescribeCapacityReservationFleets

Description

Describes one or more Capacity Reservation Fleets.

This operation returns paginated results.

Creating a Request

data DescribeCapacityReservationFleets

See: newDescribeCapacityReservationFleets smart constructor.

Constructors

DescribeCapacityReservationFleets'

Fields

capacityReservationFleetIds :: Maybe [Text]

The IDs of the Capacity Reservation Fleets to describe.

filters :: Maybe [Filter]

One or more filters.

state - The state of the Fleet (submitted | modifying | active | partially_fulfilled | expiring | expired | cancelling | cancelled | failed).

instance-match-criteria - The instance matching criteria for the Fleet. Only open is supported.

tenancy - The tenancy of the Fleet (default | dedicated).

allocation-strategy - The allocation strategy used by the Fleet. Only prioritized is supported.

nextToken :: Maybe Text

The token to use to retrieve the next page of results.

dryRun :: Maybe Bool

Checks whether you have the required permissions for the action, without actually making the request, and provides an error response. If you have the required permissions, the error response is DryRunOperation. Otherwise, it is UnauthorizedOperation.

maxResults :: Maybe Natural

The maximum number of results to return for the request in a single page. The remaining results can be seen by sending another request with the returned nextToken value. This value can be between 5 and 500. If maxResults is given a larger value than 500, you receive an error.

newDescribeCapacityReservationFleets :: DescribeCapacityReservationFleets

Create a value of DescribeCapacityReservationFleets with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:capacityReservationFleetIds:DescribeCapacityReservationFleets', describeCapacityReservationFleets_capacityReservationFleetIds - The IDs of the Capacity Reservation Fleets to describe.

$sel:filters:DescribeCapacityReservationFleets', describeCapacityReservationFleets_filters - One or more filters.

state - The state of the Fleet (submitted | modifying | active | partially_fulfilled | expiring | expired | cancelling | cancelled | failed).

instance-match-criteria - The instance matching criteria for the Fleet. Only open is supported.

tenancy - The tenancy of the Fleet (default | dedicated).

allocation-strategy - The allocation strategy used by the Fleet. Only prioritized is supported.

$sel:nextToken:DescribeCapacityReservationFleets', describeCapacityReservationFleets_nextToken - The token to use to retrieve the next page of results.

$sel:dryRun:DescribeCapacityReservationFleets', describeCapacityReservationFleets_dryRun - Checks whether you have the required permissions for the action, without actually making the request, and provides an error response. If you have the required permissions, the error response is DryRunOperation. Otherwise, it is UnauthorizedOperation.

$sel:maxResults:DescribeCapacityReservationFleets', describeCapacityReservationFleets_maxResults - The maximum number of results to return for the request in a single page. The remaining results can be seen by sending another request with the returned nextToken value. This value can be between 5 and 500. If maxResults is given a larger value than 500, you receive an error.

Request Lenses

describeCapacityReservationFleets_capacityReservationFleetIds :: Lens' DescribeCapacityReservationFleets (Maybe [Text])

The IDs of the Capacity Reservation Fleets to describe.

describeCapacityReservationFleets_filters :: Lens' DescribeCapacityReservationFleets (Maybe [Filter])

One or more filters.

state - The state of the Fleet (submitted | modifying | active | partially_fulfilled | expiring | expired | cancelling | cancelled | failed).

instance-match-criteria - The instance matching criteria for the Fleet. Only open is supported.

tenancy - The tenancy of the Fleet (default | dedicated).

allocation-strategy - The allocation strategy used by the Fleet. Only prioritized is supported.

describeCapacityReservationFleets_nextToken :: Lens' DescribeCapacityReservationFleets (Maybe Text)

The token to use to retrieve the next page of results.

describeCapacityReservationFleets_dryRun :: Lens' DescribeCapacityReservationFleets (Maybe Bool)

Checks whether you have the required permissions for the action, without actually making the request, and provides an error response. If you have the required permissions, the error response is DryRunOperation. Otherwise, it is UnauthorizedOperation.

describeCapacityReservationFleets_maxResults :: Lens' DescribeCapacityReservationFleets (Maybe Natural)

The maximum number of results to return for the request in a single page. The remaining results can be seen by sending another request with the returned nextToken value. This value can be between 5 and 500. If maxResults is given a larger value than 500, you receive an error.

Destructuring the Response

data DescribeCapacityReservationFleetsResponse

See: newDescribeCapacityReservationFleetsResponse smart constructor.

Constructors

DescribeCapacityReservationFleetsResponse'

Fields

capacityReservationFleets :: Maybe [CapacityReservationFleet]

Information about the Capacity Reservation Fleets.

nextToken :: Maybe Text

The token to use to retrieve the next page of results. This value is null when there are no more results to return.

httpStatus :: Int

The response's http status code.

newDescribeCapacityReservationFleetsResponse

Arguments

:: Int
$sel:httpStatus:DescribeCapacityReservationFleetsResponse'

-> DescribeCapacityReservationFleetsResponse

Create a value of DescribeCapacityReservationFleetsResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:capacityReservationFleets:DescribeCapacityReservationFleetsResponse', describeCapacityReservationFleetsResponse_capacityReservationFleets - Information about the Capacity Reservation Fleets.

$sel:nextToken:DescribeCapacityReservationFleets', describeCapacityReservationFleetsResponse_nextToken - The token to use to retrieve the next page of results. This value is null when there are no more results to return.

$sel:httpStatus:DescribeCapacityReservationFleetsResponse', describeCapacityReservationFleetsResponse_httpStatus - The response's http status code.

Response Lenses

describeCapacityReservationFleetsResponse_capacityReservationFleets :: Lens' DescribeCapacityReservationFleetsResponse (Maybe [CapacityReservationFleet])

Information about the Capacity Reservation Fleets.

describeCapacityReservationFleetsResponse_nextToken :: Lens' DescribeCapacityReservationFleetsResponse (Maybe Text)

The token to use to retrieve the next page of results. This value is null when there are no more results to return.

describeCapacityReservationFleetsResponse_httpStatus :: Lens' DescribeCapacityReservationFleetsResponse Int

The response's http status code.