Amazonka.EC2.DescribeElasticGpus

Description

Describes the Elastic Graphics accelerator associated with your instances. For more information about Elastic Graphics, see Amazon Elastic Graphics.

Creating a Request

data DescribeElasticGpus

See: newDescribeElasticGpus smart constructor.

Constructors

DescribeElasticGpus'

Fields

filters :: Maybe [Filter]

The filters.

availability-zone - The Availability Zone in which the Elastic Graphics accelerator resides.

elastic-gpu-health - The status of the Elastic Graphics accelerator (OK | IMPAIRED).

elastic-gpu-state - The state of the Elastic Graphics accelerator (ATTACHED).

elastic-gpu-type - The type of Elastic Graphics accelerator; for example, eg1.medium.

instance-id - The ID of the instance to which the Elastic Graphics accelerator is associated.

nextToken :: Maybe Text

The token to request the next page of results.

dryRun :: Maybe Bool

Checks whether you have the required permissions for the action, without actually making the request, and provides an error response. If you have the required permissions, the error response is DryRunOperation. Otherwise, it is UnauthorizedOperation.

maxResults :: Maybe Natural

The maximum number of results to return in a single call. To retrieve the remaining results, make another call with the returned NextToken value. This value can be between 5 and 1000.

elasticGpuIds :: Maybe [Text]

The Elastic Graphics accelerator IDs.

newDescribeElasticGpus :: DescribeElasticGpus

Create a value of DescribeElasticGpus with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:filters:DescribeElasticGpus', describeElasticGpus_filters - The filters.

availability-zone - The Availability Zone in which the Elastic Graphics accelerator resides.

elastic-gpu-health - The status of the Elastic Graphics accelerator (OK | IMPAIRED).

elastic-gpu-state - The state of the Elastic Graphics accelerator (ATTACHED).

elastic-gpu-type - The type of Elastic Graphics accelerator; for example, eg1.medium.

instance-id - The ID of the instance to which the Elastic Graphics accelerator is associated.

$sel:nextToken:DescribeElasticGpus', describeElasticGpus_nextToken - The token to request the next page of results.

$sel:dryRun:DescribeElasticGpus', describeElasticGpus_dryRun - Checks whether you have the required permissions for the action, without actually making the request, and provides an error response. If you have the required permissions, the error response is DryRunOperation. Otherwise, it is UnauthorizedOperation.

$sel:maxResults:DescribeElasticGpus', describeElasticGpus_maxResults - The maximum number of results to return in a single call. To retrieve the remaining results, make another call with the returned NextToken value. This value can be between 5 and 1000.

$sel:elasticGpuIds:DescribeElasticGpus', describeElasticGpus_elasticGpuIds - The Elastic Graphics accelerator IDs.

Request Lenses

describeElasticGpus_filters :: Lens' DescribeElasticGpus (Maybe [Filter])

The filters.

availability-zone - The Availability Zone in which the Elastic Graphics accelerator resides.

elastic-gpu-health - The status of the Elastic Graphics accelerator (OK | IMPAIRED).

elastic-gpu-state - The state of the Elastic Graphics accelerator (ATTACHED).

elastic-gpu-type - The type of Elastic Graphics accelerator; for example, eg1.medium.

instance-id - The ID of the instance to which the Elastic Graphics accelerator is associated.

describeElasticGpus_nextToken :: Lens' DescribeElasticGpus (Maybe Text)

The token to request the next page of results.

describeElasticGpus_dryRun :: Lens' DescribeElasticGpus (Maybe Bool)

Checks whether you have the required permissions for the action, without actually making the request, and provides an error response. If you have the required permissions, the error response is DryRunOperation. Otherwise, it is UnauthorizedOperation.

describeElasticGpus_maxResults :: Lens' DescribeElasticGpus (Maybe Natural)

The maximum number of results to return in a single call. To retrieve the remaining results, make another call with the returned NextToken value. This value can be between 5 and 1000.

describeElasticGpus_elasticGpuIds :: Lens' DescribeElasticGpus (Maybe [Text])

The Elastic Graphics accelerator IDs.

Destructuring the Response

data DescribeElasticGpusResponse

See: newDescribeElasticGpusResponse smart constructor.

Constructors

DescribeElasticGpusResponse'

Fields

elasticGpuSet :: Maybe [ElasticGpus]

Information about the Elastic Graphics accelerators.

nextToken :: Maybe Text

The token to use to retrieve the next page of results. This value is null when there are no more results to return.

maxResults :: Maybe Int

The total number of items to return. If the total number of items available is more than the value specified in max-items then a Next-Token will be provided in the output that you can use to resume pagination.

httpStatus :: Int

The response's http status code.

newDescribeElasticGpusResponse

Arguments

:: Int
$sel:httpStatus:DescribeElasticGpusResponse'

-> DescribeElasticGpusResponse

Create a value of DescribeElasticGpusResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:elasticGpuSet:DescribeElasticGpusResponse', describeElasticGpusResponse_elasticGpuSet - Information about the Elastic Graphics accelerators.

$sel:nextToken:DescribeElasticGpus', describeElasticGpusResponse_nextToken - The token to use to retrieve the next page of results. This value is null when there are no more results to return.

$sel:maxResults:DescribeElasticGpus', describeElasticGpusResponse_maxResults - The total number of items to return. If the total number of items available is more than the value specified in max-items then a Next-Token will be provided in the output that you can use to resume pagination.

$sel:httpStatus:DescribeElasticGpusResponse', describeElasticGpusResponse_httpStatus - The response's http status code.

Response Lenses

describeElasticGpusResponse_elasticGpuSet :: Lens' DescribeElasticGpusResponse (Maybe [ElasticGpus])

Information about the Elastic Graphics accelerators.

describeElasticGpusResponse_nextToken :: Lens' DescribeElasticGpusResponse (Maybe Text)

The token to use to retrieve the next page of results. This value is null when there are no more results to return.

describeElasticGpusResponse_maxResults :: Lens' DescribeElasticGpusResponse (Maybe Int)

The total number of items to return. If the total number of items available is more than the value specified in max-items then a Next-Token will be provided in the output that you can use to resume pagination.

describeElasticGpusResponse_httpStatus :: Lens' DescribeElasticGpusResponse Int

The response's http status code.