Amazonka.EC2.DescribeHostReservationOfferings

Description

Describes the Dedicated Host reservations that are available to purchase.

The results describe all of the Dedicated Host reservation offerings, including offerings that might not match the instance family and Region of your Dedicated Hosts. When purchasing an offering, ensure that the instance family and Region of the offering matches that of the Dedicated Hosts with which it is to be associated. For more information about supported instance types, see Dedicated Hosts in the Amazon EC2 User Guide.

This operation returns paginated results.

Creating a Request

data DescribeHostReservationOfferings

See: newDescribeHostReservationOfferings smart constructor.

Constructors

DescribeHostReservationOfferings'

Fields

maxDuration :: Maybe Int

This is the maximum duration of the reservation to purchase, specified in seconds. Reservations are available in one-year and three-year terms. The number of seconds specified must be the number of seconds in a year (365x24x60x60) times one of the supported durations (1 or 3). For example, specify 94608000 for three years.

nextToken :: Maybe Text

The token to use to retrieve the next page of results.

minDuration :: Maybe Int

This is the minimum duration of the reservation you'd like to purchase, specified in seconds. Reservations are available in one-year and three-year terms. The number of seconds specified must be the number of seconds in a year (365x24x60x60) times one of the supported durations (1 or 3). For example, specify 31536000 for one year.

offeringId :: Maybe Text

The ID of the reservation offering.

filter' :: Maybe [Filter]

The filters.

instance-family - The instance family of the offering (for example, m4).

payment-option - The payment option (NoUpfront | PartialUpfront | AllUpfront).

maxResults :: Maybe Natural

The maximum number of results to return for the request in a single page. The remaining results can be seen by sending another request with the returned nextToken value. This value can be between 5 and 500. If maxResults is given a larger value than 500, you receive an error.

newDescribeHostReservationOfferings :: DescribeHostReservationOfferings

Create a value of DescribeHostReservationOfferings with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:maxDuration:DescribeHostReservationOfferings', describeHostReservationOfferings_maxDuration - This is the maximum duration of the reservation to purchase, specified in seconds. Reservations are available in one-year and three-year terms. The number of seconds specified must be the number of seconds in a year (365x24x60x60) times one of the supported durations (1 or 3). For example, specify 94608000 for three years.

$sel:nextToken:DescribeHostReservationOfferings', describeHostReservationOfferings_nextToken - The token to use to retrieve the next page of results.

$sel:minDuration:DescribeHostReservationOfferings', describeHostReservationOfferings_minDuration - This is the minimum duration of the reservation you'd like to purchase, specified in seconds. Reservations are available in one-year and three-year terms. The number of seconds specified must be the number of seconds in a year (365x24x60x60) times one of the supported durations (1 or 3). For example, specify 31536000 for one year.

$sel:offeringId:DescribeHostReservationOfferings', describeHostReservationOfferings_offeringId - The ID of the reservation offering.

$sel:filter':DescribeHostReservationOfferings', describeHostReservationOfferings_filter - The filters.

instance-family - The instance family of the offering (for example, m4).

payment-option - The payment option (NoUpfront | PartialUpfront | AllUpfront).

$sel:maxResults:DescribeHostReservationOfferings', describeHostReservationOfferings_maxResults - The maximum number of results to return for the request in a single page. The remaining results can be seen by sending another request with the returned nextToken value. This value can be between 5 and 500. If maxResults is given a larger value than 500, you receive an error.

Request Lenses

describeHostReservationOfferings_maxDuration :: Lens' DescribeHostReservationOfferings (Maybe Int)

This is the maximum duration of the reservation to purchase, specified in seconds. Reservations are available in one-year and three-year terms. The number of seconds specified must be the number of seconds in a year (365x24x60x60) times one of the supported durations (1 or 3). For example, specify 94608000 for three years.

describeHostReservationOfferings_nextToken :: Lens' DescribeHostReservationOfferings (Maybe Text)

The token to use to retrieve the next page of results.

describeHostReservationOfferings_minDuration :: Lens' DescribeHostReservationOfferings (Maybe Int)

This is the minimum duration of the reservation you'd like to purchase, specified in seconds. Reservations are available in one-year and three-year terms. The number of seconds specified must be the number of seconds in a year (365x24x60x60) times one of the supported durations (1 or 3). For example, specify 31536000 for one year.

describeHostReservationOfferings_offeringId :: Lens' DescribeHostReservationOfferings (Maybe Text)

The ID of the reservation offering.

describeHostReservationOfferings_filter :: Lens' DescribeHostReservationOfferings (Maybe [Filter])

The filters.

instance-family - The instance family of the offering (for example, m4).

payment-option - The payment option (NoUpfront | PartialUpfront | AllUpfront).

describeHostReservationOfferings_maxResults :: Lens' DescribeHostReservationOfferings (Maybe Natural)

The maximum number of results to return for the request in a single page. The remaining results can be seen by sending another request with the returned nextToken value. This value can be between 5 and 500. If maxResults is given a larger value than 500, you receive an error.

Destructuring the Response

data DescribeHostReservationOfferingsResponse

See: newDescribeHostReservationOfferingsResponse smart constructor.

Constructors

DescribeHostReservationOfferingsResponse'

Fields

offeringSet :: Maybe [HostOffering]

Information about the offerings.

nextToken :: Maybe Text

The token to use to retrieve the next page of results. This value is null when there are no more results to return.

httpStatus :: Int

The response's http status code.

newDescribeHostReservationOfferingsResponse

Arguments

:: Int
$sel:httpStatus:DescribeHostReservationOfferingsResponse'

-> DescribeHostReservationOfferingsResponse

Create a value of DescribeHostReservationOfferingsResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:offeringSet:DescribeHostReservationOfferingsResponse', describeHostReservationOfferingsResponse_offeringSet - Information about the offerings.

$sel:nextToken:DescribeHostReservationOfferings', describeHostReservationOfferingsResponse_nextToken - The token to use to retrieve the next page of results. This value is null when there are no more results to return.

$sel:httpStatus:DescribeHostReservationOfferingsResponse', describeHostReservationOfferingsResponse_httpStatus - The response's http status code.

Response Lenses

describeHostReservationOfferingsResponse_offeringSet :: Lens' DescribeHostReservationOfferingsResponse (Maybe [HostOffering])

Information about the offerings.

describeHostReservationOfferingsResponse_nextToken :: Lens' DescribeHostReservationOfferingsResponse (Maybe Text)

The token to use to retrieve the next page of results. This value is null when there are no more results to return.

describeHostReservationOfferingsResponse_httpStatus :: Lens' DescribeHostReservationOfferingsResponse Int

The response's http status code.