Amazonka.EC2.DescribeIamInstanceProfileAssociations

Description

Describes your IAM instance profile associations.

This operation returns paginated results.

Creating a Request

data DescribeIamInstanceProfileAssociations

See: newDescribeIamInstanceProfileAssociations smart constructor.

Constructors

DescribeIamInstanceProfileAssociations'

Fields

filters :: Maybe [Filter]

The filters.

instance-id - The ID of the instance.

state - The state of the association (associating | associated | disassociating).

nextToken :: Maybe Text

The token to request the next page of results.

associationIds :: Maybe [Text]

The IAM instance profile associations.

maxResults :: Maybe Natural

The maximum number of results to return in a single call. To retrieve the remaining results, make another call with the returned NextToken value.

newDescribeIamInstanceProfileAssociations :: DescribeIamInstanceProfileAssociations

Create a value of DescribeIamInstanceProfileAssociations with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:filters:DescribeIamInstanceProfileAssociations', describeIamInstanceProfileAssociations_filters - The filters.

instance-id - The ID of the instance.

state - The state of the association (associating | associated | disassociating).

$sel:nextToken:DescribeIamInstanceProfileAssociations', describeIamInstanceProfileAssociations_nextToken - The token to request the next page of results.

$sel:associationIds:DescribeIamInstanceProfileAssociations', describeIamInstanceProfileAssociations_associationIds - The IAM instance profile associations.

$sel:maxResults:DescribeIamInstanceProfileAssociations', describeIamInstanceProfileAssociations_maxResults - The maximum number of results to return in a single call. To retrieve the remaining results, make another call with the returned NextToken value.

Request Lenses

describeIamInstanceProfileAssociations_filters :: Lens' DescribeIamInstanceProfileAssociations (Maybe [Filter])

The filters.

instance-id - The ID of the instance.

state - The state of the association (associating | associated | disassociating).

describeIamInstanceProfileAssociations_nextToken :: Lens' DescribeIamInstanceProfileAssociations (Maybe Text)

The token to request the next page of results.

describeIamInstanceProfileAssociations_associationIds :: Lens' DescribeIamInstanceProfileAssociations (Maybe [Text])

The IAM instance profile associations.

describeIamInstanceProfileAssociations_maxResults :: Lens' DescribeIamInstanceProfileAssociations (Maybe Natural)

The maximum number of results to return in a single call. To retrieve the remaining results, make another call with the returned NextToken value.

Destructuring the Response

data DescribeIamInstanceProfileAssociationsResponse

See: newDescribeIamInstanceProfileAssociationsResponse smart constructor.

Constructors

DescribeIamInstanceProfileAssociationsResponse'

Fields

iamInstanceProfileAssociations :: Maybe [IamInstanceProfileAssociation]

Information about the IAM instance profile associations.

nextToken :: Maybe Text

The token to use to retrieve the next page of results. This value is null when there are no more results to return.

httpStatus :: Int

The response's http status code.

newDescribeIamInstanceProfileAssociationsResponse

Arguments

:: Int
$sel:httpStatus:DescribeIamInstanceProfileAssociationsResponse'

-> DescribeIamInstanceProfileAssociationsResponse

Create a value of DescribeIamInstanceProfileAssociationsResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:iamInstanceProfileAssociations:DescribeIamInstanceProfileAssociationsResponse', describeIamInstanceProfileAssociationsResponse_iamInstanceProfileAssociations - Information about the IAM instance profile associations.

$sel:nextToken:DescribeIamInstanceProfileAssociations', describeIamInstanceProfileAssociationsResponse_nextToken - The token to use to retrieve the next page of results. This value is null when there are no more results to return.

$sel:httpStatus:DescribeIamInstanceProfileAssociationsResponse', describeIamInstanceProfileAssociationsResponse_httpStatus - The response's http status code.

Response Lenses

describeIamInstanceProfileAssociationsResponse_iamInstanceProfileAssociations :: Lens' DescribeIamInstanceProfileAssociationsResponse (Maybe [IamInstanceProfileAssociation])

Information about the IAM instance profile associations.

describeIamInstanceProfileAssociationsResponse_nextToken :: Lens' DescribeIamInstanceProfileAssociationsResponse (Maybe Text)

The token to use to retrieve the next page of results. This value is null when there are no more results to return.

describeIamInstanceProfileAssociationsResponse_httpStatus :: Lens' DescribeIamInstanceProfileAssociationsResponse Int

The response's http status code.