Amazonka.EC2.DescribeInstanceAttribute

Description

Describes the specified attribute of the specified instance. You can specify only one attribute at a time. Valid attribute values are: instanceType | kernel | ramdisk | userData | disableApiTermination | instanceInitiatedShutdownBehavior | rootDeviceName | blockDeviceMapping | productCodes | sourceDestCheck | groupSet | ebsOptimized | sriovNetSupport

Creating a Request

data DescribeInstanceAttribute

See: newDescribeInstanceAttribute smart constructor.

Constructors

DescribeInstanceAttribute'

Fields

dryRun :: Maybe Bool

Checks whether you have the required permissions for the action, without actually making the request, and provides an error response. If you have the required permissions, the error response is DryRunOperation. Otherwise, it is UnauthorizedOperation.

attribute :: InstanceAttributeName

The instance attribute.

Note: The enaSupport attribute is not supported at this time.

instanceId :: Text

The ID of the instance.

newDescribeInstanceAttribute

Arguments

:: InstanceAttributeName
$sel:attribute:DescribeInstanceAttribute'

-> Text
$sel:instanceId:DescribeInstanceAttribute'

-> DescribeInstanceAttribute

Create a value of DescribeInstanceAttribute with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:dryRun:DescribeInstanceAttribute', describeInstanceAttribute_dryRun - Checks whether you have the required permissions for the action, without actually making the request, and provides an error response. If you have the required permissions, the error response is DryRunOperation. Otherwise, it is UnauthorizedOperation.

$sel:attribute:DescribeInstanceAttribute', describeInstanceAttribute_attribute - The instance attribute.

Note: The enaSupport attribute is not supported at this time.

$sel:instanceId:DescribeInstanceAttribute', describeInstanceAttribute_instanceId - The ID of the instance.

Request Lenses

describeInstanceAttribute_dryRun :: Lens' DescribeInstanceAttribute (Maybe Bool)

Checks whether you have the required permissions for the action, without actually making the request, and provides an error response. If you have the required permissions, the error response is DryRunOperation. Otherwise, it is UnauthorizedOperation.

describeInstanceAttribute_attribute :: Lens' DescribeInstanceAttribute InstanceAttributeName

The instance attribute.

Note: The enaSupport attribute is not supported at this time.

describeInstanceAttribute_instanceId :: Lens' DescribeInstanceAttribute Text

The ID of the instance.

Destructuring the Response

data DescribeInstanceAttributeResponse

Describes an instance attribute.

See: newDescribeInstanceAttributeResponse smart constructor.

Constructors

DescribeInstanceAttributeResponse'

Fields

instanceId :: Maybe Text

The ID of the instance.

groups :: Maybe [GroupIdentifier]

The security groups associated with the instance.

enaSupport :: Maybe AttributeBooleanValue

Indicates whether enhanced networking with ENA is enabled.

sourceDestCheck :: Maybe AttributeBooleanValue

Enable or disable source/destination checks, which ensure that the instance is either the source or the destination of any traffic that it receives. If the value is true, source/destination checks are enabled; otherwise, they are disabled. The default value is true. You must disable source/destination checks if the instance runs services such as network address translation, routing, or firewalls.

disableApiTermination :: Maybe AttributeBooleanValue

If the value is true, you can't terminate the instance through the Amazon EC2 console, CLI, or API; otherwise, you can.

enclaveOptions :: Maybe EnclaveOptions

To enable the instance for Amazon Web Services Nitro Enclaves, set this parameter to true; otherwise, set it to false.

ramdiskId :: Maybe AttributeValue

The RAM disk ID.

kernelId :: Maybe AttributeValue

The kernel ID.

rootDeviceName :: Maybe AttributeValue

The device name of the root device volume (for example, /dev/sda1).

instanceType :: Maybe AttributeValue

The instance type.

sriovNetSupport :: Maybe AttributeValue

Indicates whether enhanced networking with the Intel 82599 Virtual Function interface is enabled.

ebsOptimized :: Maybe AttributeBooleanValue

Indicates whether the instance is optimized for Amazon EBS I/O.

userData :: Maybe AttributeValue

The user data.

instanceInitiatedShutdownBehavior :: Maybe AttributeValue

Indicates whether an instance stops or terminates when you initiate shutdown from the instance (using the operating system command for system shutdown).

productCodes :: Maybe [ProductCode]

A list of product codes.

blockDeviceMappings :: Maybe [InstanceBlockDeviceMapping]

The block device mapping of the instance.

httpStatus :: Int

The response's http status code.

newDescribeInstanceAttributeResponse

Arguments

:: Int
$sel:httpStatus:DescribeInstanceAttributeResponse'

-> DescribeInstanceAttributeResponse

Create a value of DescribeInstanceAttributeResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:instanceId:DescribeInstanceAttribute', describeInstanceAttributeResponse_instanceId - The ID of the instance.

$sel:groups:DescribeInstanceAttributeResponse', describeInstanceAttributeResponse_groups - The security groups associated with the instance.

$sel:enaSupport:DescribeInstanceAttributeResponse', describeInstanceAttributeResponse_enaSupport - Indicates whether enhanced networking with ENA is enabled.

$sel:sourceDestCheck:DescribeInstanceAttributeResponse', describeInstanceAttributeResponse_sourceDestCheck - Enable or disable source/destination checks, which ensure that the instance is either the source or the destination of any traffic that it receives. If the value is true, source/destination checks are enabled; otherwise, they are disabled. The default value is true. You must disable source/destination checks if the instance runs services such as network address translation, routing, or firewalls.

$sel:disableApiTermination:DescribeInstanceAttributeResponse', describeInstanceAttributeResponse_disableApiTermination - If the value is true, you can't terminate the instance through the Amazon EC2 console, CLI, or API; otherwise, you can.

$sel:enclaveOptions:DescribeInstanceAttributeResponse', describeInstanceAttributeResponse_enclaveOptions - To enable the instance for Amazon Web Services Nitro Enclaves, set this parameter to true; otherwise, set it to false.

$sel:ramdiskId:DescribeInstanceAttributeResponse', describeInstanceAttributeResponse_ramdiskId - The RAM disk ID.

$sel:kernelId:DescribeInstanceAttributeResponse', describeInstanceAttributeResponse_kernelId - The kernel ID.

$sel:rootDeviceName:DescribeInstanceAttributeResponse', describeInstanceAttributeResponse_rootDeviceName - The device name of the root device volume (for example, /dev/sda1).

$sel:instanceType:DescribeInstanceAttributeResponse', describeInstanceAttributeResponse_instanceType - The instance type.

$sel:sriovNetSupport:DescribeInstanceAttributeResponse', describeInstanceAttributeResponse_sriovNetSupport - Indicates whether enhanced networking with the Intel 82599 Virtual Function interface is enabled.

$sel:ebsOptimized:DescribeInstanceAttributeResponse', describeInstanceAttributeResponse_ebsOptimized - Indicates whether the instance is optimized for Amazon EBS I/O.

$sel:userData:DescribeInstanceAttributeResponse', describeInstanceAttributeResponse_userData - The user data.

$sel:instanceInitiatedShutdownBehavior:DescribeInstanceAttributeResponse', describeInstanceAttributeResponse_instanceInitiatedShutdownBehavior - Indicates whether an instance stops or terminates when you initiate shutdown from the instance (using the operating system command for system shutdown).

$sel:productCodes:DescribeInstanceAttributeResponse', describeInstanceAttributeResponse_productCodes - A list of product codes.

$sel:blockDeviceMappings:DescribeInstanceAttributeResponse', describeInstanceAttributeResponse_blockDeviceMappings - The block device mapping of the instance.

$sel:httpStatus:DescribeInstanceAttributeResponse', describeInstanceAttributeResponse_httpStatus - The response's http status code.

Response Lenses

describeInstanceAttributeResponse_instanceId :: Lens' DescribeInstanceAttributeResponse (Maybe Text)

The ID of the instance.

describeInstanceAttributeResponse_groups :: Lens' DescribeInstanceAttributeResponse (Maybe [GroupIdentifier])

The security groups associated with the instance.

describeInstanceAttributeResponse_enaSupport :: Lens' DescribeInstanceAttributeResponse (Maybe AttributeBooleanValue)

Indicates whether enhanced networking with ENA is enabled.

describeInstanceAttributeResponse_sourceDestCheck :: Lens' DescribeInstanceAttributeResponse (Maybe AttributeBooleanValue)

Enable or disable source/destination checks, which ensure that the instance is either the source or the destination of any traffic that it receives. If the value is true, source/destination checks are enabled; otherwise, they are disabled. The default value is true. You must disable source/destination checks if the instance runs services such as network address translation, routing, or firewalls.

describeInstanceAttributeResponse_disableApiTermination :: Lens' DescribeInstanceAttributeResponse (Maybe AttributeBooleanValue)

If the value is true, you can't terminate the instance through the Amazon EC2 console, CLI, or API; otherwise, you can.

describeInstanceAttributeResponse_enclaveOptions :: Lens' DescribeInstanceAttributeResponse (Maybe EnclaveOptions)

To enable the instance for Amazon Web Services Nitro Enclaves, set this parameter to true; otherwise, set it to false.

describeInstanceAttributeResponse_ramdiskId :: Lens' DescribeInstanceAttributeResponse (Maybe AttributeValue)

The RAM disk ID.

describeInstanceAttributeResponse_kernelId :: Lens' DescribeInstanceAttributeResponse (Maybe AttributeValue)

The kernel ID.

describeInstanceAttributeResponse_rootDeviceName :: Lens' DescribeInstanceAttributeResponse (Maybe AttributeValue)

The device name of the root device volume (for example, /dev/sda1).

describeInstanceAttributeResponse_instanceType :: Lens' DescribeInstanceAttributeResponse (Maybe AttributeValue)

The instance type.

describeInstanceAttributeResponse_sriovNetSupport :: Lens' DescribeInstanceAttributeResponse (Maybe AttributeValue)

Indicates whether enhanced networking with the Intel 82599 Virtual Function interface is enabled.

describeInstanceAttributeResponse_ebsOptimized :: Lens' DescribeInstanceAttributeResponse (Maybe AttributeBooleanValue)

Indicates whether the instance is optimized for Amazon EBS I/O.

describeInstanceAttributeResponse_userData :: Lens' DescribeInstanceAttributeResponse (Maybe AttributeValue)

The user data.

describeInstanceAttributeResponse_instanceInitiatedShutdownBehavior :: Lens' DescribeInstanceAttributeResponse (Maybe AttributeValue)

Indicates whether an instance stops or terminates when you initiate shutdown from the instance (using the operating system command for system shutdown).

describeInstanceAttributeResponse_productCodes :: Lens' DescribeInstanceAttributeResponse (Maybe [ProductCode])

A list of product codes.

describeInstanceAttributeResponse_blockDeviceMappings :: Lens' DescribeInstanceAttributeResponse (Maybe [InstanceBlockDeviceMapping])

The block device mapping of the instance.

describeInstanceAttributeResponse_httpStatus :: Lens' DescribeInstanceAttributeResponse Int

The response's http status code.