Amazonka.EC2.DescribeLocalGatewayRouteTableVirtualInterfaceGroupAssociations

Description

Describes the associations between virtual interface groups and local gateway route tables.

This operation returns paginated results.

Creating a Request

data DescribeLocalGatewayRouteTableVirtualInterfaceGroupAssociations

See: newDescribeLocalGatewayRouteTableVirtualInterfaceGroupAssociations smart constructor.

Constructors

DescribeLocalGatewayRouteTableVirtualInterfaceGroupAssociations'

Fields

filters :: Maybe [Filter]

One or more filters.

local-gateway-id - The ID of a local gateway.

local-gateway-route-table-id - The ID of the local gateway route table.

local-gateway-route-table-virtual-interface-group-association-id - The ID of the association.

local-gateway-route-table-virtual-interface-group-id - The ID of the virtual interface group.

state - The state of the association.

localGatewayRouteTableVirtualInterfaceGroupAssociationIds :: Maybe [Text]

The IDs of the associations.

nextToken :: Maybe Text

The token for the next page of results.

dryRun :: Maybe Bool

Checks whether you have the required permissions for the action, without actually making the request, and provides an error response. If you have the required permissions, the error response is DryRunOperation. Otherwise, it is UnauthorizedOperation.

maxResults :: Maybe Natural

The maximum number of results to return with a single call. To retrieve the remaining results, make another call with the returned nextToken value.

newDescribeLocalGatewayRouteTableVirtualInterfaceGroupAssociations :: DescribeLocalGatewayRouteTableVirtualInterfaceGroupAssociations

Create a value of DescribeLocalGatewayRouteTableVirtualInterfaceGroupAssociations with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:filters:DescribeLocalGatewayRouteTableVirtualInterfaceGroupAssociations', describeLocalGatewayRouteTableVirtualInterfaceGroupAssociations_filters - One or more filters.

local-gateway-id - The ID of a local gateway.

local-gateway-route-table-id - The ID of the local gateway route table.

local-gateway-route-table-virtual-interface-group-association-id - The ID of the association.

local-gateway-route-table-virtual-interface-group-id - The ID of the virtual interface group.

state - The state of the association.

$sel:localGatewayRouteTableVirtualInterfaceGroupAssociationIds:DescribeLocalGatewayRouteTableVirtualInterfaceGroupAssociations', describeLocalGatewayRouteTableVirtualInterfaceGroupAssociations_localGatewayRouteTableVirtualInterfaceGroupAssociationIds - The IDs of the associations.

$sel:nextToken:DescribeLocalGatewayRouteTableVirtualInterfaceGroupAssociations', describeLocalGatewayRouteTableVirtualInterfaceGroupAssociations_nextToken - The token for the next page of results.

$sel:dryRun:DescribeLocalGatewayRouteTableVirtualInterfaceGroupAssociations', describeLocalGatewayRouteTableVirtualInterfaceGroupAssociations_dryRun - Checks whether you have the required permissions for the action, without actually making the request, and provides an error response. If you have the required permissions, the error response is DryRunOperation. Otherwise, it is UnauthorizedOperation.

$sel:maxResults:DescribeLocalGatewayRouteTableVirtualInterfaceGroupAssociations', describeLocalGatewayRouteTableVirtualInterfaceGroupAssociations_maxResults - The maximum number of results to return with a single call. To retrieve the remaining results, make another call with the returned nextToken value.

Request Lenses

describeLocalGatewayRouteTableVirtualInterfaceGroupAssociations_filters :: Lens' DescribeLocalGatewayRouteTableVirtualInterfaceGroupAssociations (Maybe [Filter])

One or more filters.

local-gateway-id - The ID of a local gateway.

local-gateway-route-table-id - The ID of the local gateway route table.

local-gateway-route-table-virtual-interface-group-association-id - The ID of the association.

local-gateway-route-table-virtual-interface-group-id - The ID of the virtual interface group.

state - The state of the association.

describeLocalGatewayRouteTableVirtualInterfaceGroupAssociations_localGatewayRouteTableVirtualInterfaceGroupAssociationIds :: Lens' DescribeLocalGatewayRouteTableVirtualInterfaceGroupAssociations (Maybe [Text])

The IDs of the associations.

describeLocalGatewayRouteTableVirtualInterfaceGroupAssociations_nextToken :: Lens' DescribeLocalGatewayRouteTableVirtualInterfaceGroupAssociations (Maybe Text)

The token for the next page of results.

describeLocalGatewayRouteTableVirtualInterfaceGroupAssociations_dryRun :: Lens' DescribeLocalGatewayRouteTableVirtualInterfaceGroupAssociations (Maybe Bool)

Checks whether you have the required permissions for the action, without actually making the request, and provides an error response. If you have the required permissions, the error response is DryRunOperation. Otherwise, it is UnauthorizedOperation.

describeLocalGatewayRouteTableVirtualInterfaceGroupAssociations_maxResults :: Lens' DescribeLocalGatewayRouteTableVirtualInterfaceGroupAssociations (Maybe Natural)

The maximum number of results to return with a single call. To retrieve the remaining results, make another call with the returned nextToken value.

Destructuring the Response

data DescribeLocalGatewayRouteTableVirtualInterfaceGroupAssociationsResponse

See: newDescribeLocalGatewayRouteTableVirtualInterfaceGroupAssociationsResponse smart constructor.

Constructors

DescribeLocalGatewayRouteTableVirtualInterfaceGroupAssociationsResponse'

Fields

nextToken :: Maybe Text

The token to use to retrieve the next page of results. This value is null when there are no more results to return.

localGatewayRouteTableVirtualInterfaceGroupAssociations :: Maybe [LocalGatewayRouteTableVirtualInterfaceGroupAssociation]

Information about the associations.

httpStatus :: Int

The response's http status code.

newDescribeLocalGatewayRouteTableVirtualInterfaceGroupAssociationsResponse

Arguments

:: Int
$sel:httpStatus:DescribeLocalGatewayRouteTableVirtualInterfaceGroupAssociationsResponse'

-> DescribeLocalGatewayRouteTableVirtualInterfaceGroupAssociationsResponse

Create a value of DescribeLocalGatewayRouteTableVirtualInterfaceGroupAssociationsResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:nextToken:DescribeLocalGatewayRouteTableVirtualInterfaceGroupAssociations', describeLocalGatewayRouteTableVirtualInterfaceGroupAssociationsResponse_nextToken - The token to use to retrieve the next page of results. This value is null when there are no more results to return.

$sel:localGatewayRouteTableVirtualInterfaceGroupAssociations:DescribeLocalGatewayRouteTableVirtualInterfaceGroupAssociationsResponse', describeLocalGatewayRouteTableVirtualInterfaceGroupAssociationsResponse_localGatewayRouteTableVirtualInterfaceGroupAssociations - Information about the associations.

$sel:httpStatus:DescribeLocalGatewayRouteTableVirtualInterfaceGroupAssociationsResponse', describeLocalGatewayRouteTableVirtualInterfaceGroupAssociationsResponse_httpStatus - The response's http status code.

Response Lenses

describeLocalGatewayRouteTableVirtualInterfaceGroupAssociationsResponse_nextToken :: Lens' DescribeLocalGatewayRouteTableVirtualInterfaceGroupAssociationsResponse (Maybe Text)

The token to use to retrieve the next page of results. This value is null when there are no more results to return.

describeLocalGatewayRouteTableVirtualInterfaceGroupAssociationsResponse_localGatewayRouteTableVirtualInterfaceGroupAssociations :: Lens' DescribeLocalGatewayRouteTableVirtualInterfaceGroupAssociationsResponse (Maybe [LocalGatewayRouteTableVirtualInterfaceGroupAssociation])

Information about the associations.

describeLocalGatewayRouteTableVirtualInterfaceGroupAssociationsResponse_httpStatus :: Lens' DescribeLocalGatewayRouteTableVirtualInterfaceGroupAssociationsResponse Int

The response's http status code.