Amazonka.EC2.DescribeMovingAddresses

Description

Describes your Elastic IP addresses that are being moved to the EC2-VPC platform, or that are being restored to the EC2-Classic platform. This request does not return information about any other Elastic IP addresses in your account.

This operation returns paginated results.

Creating a Request

data DescribeMovingAddresses

See: newDescribeMovingAddresses smart constructor.

Constructors

DescribeMovingAddresses'

Fields

filters :: Maybe [Filter]

One or more filters.

moving-status - The status of the Elastic IP address (MovingToVpc | RestoringToClassic).

publicIps :: Maybe [Text]

One or more Elastic IP addresses.

nextToken :: Maybe Text

The token for the next page of results.

dryRun :: Maybe Bool

Checks whether you have the required permissions for the action, without actually making the request, and provides an error response. If you have the required permissions, the error response is DryRunOperation. Otherwise, it is UnauthorizedOperation.

maxResults :: Maybe Natural

The maximum number of results to return for the request in a single page. The remaining results of the initial request can be seen by sending another request with the returned NextToken value. This value can be between 5 and 1000; if MaxResults is given a value outside of this range, an error is returned.

Default: If no value is provided, the default is 1000.

newDescribeMovingAddresses :: DescribeMovingAddresses

Create a value of DescribeMovingAddresses with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:filters:DescribeMovingAddresses', describeMovingAddresses_filters - One or more filters.

moving-status - The status of the Elastic IP address (MovingToVpc | RestoringToClassic).

$sel:publicIps:DescribeMovingAddresses', describeMovingAddresses_publicIps - One or more Elastic IP addresses.

$sel:nextToken:DescribeMovingAddresses', describeMovingAddresses_nextToken - The token for the next page of results.

$sel:dryRun:DescribeMovingAddresses', describeMovingAddresses_dryRun - Checks whether you have the required permissions for the action, without actually making the request, and provides an error response. If you have the required permissions, the error response is DryRunOperation. Otherwise, it is UnauthorizedOperation.

$sel:maxResults:DescribeMovingAddresses', describeMovingAddresses_maxResults - The maximum number of results to return for the request in a single page. The remaining results of the initial request can be seen by sending another request with the returned NextToken value. This value can be between 5 and 1000; if MaxResults is given a value outside of this range, an error is returned.

Default: If no value is provided, the default is 1000.

Request Lenses

describeMovingAddresses_filters :: Lens' DescribeMovingAddresses (Maybe [Filter])

One or more filters.

moving-status - The status of the Elastic IP address (MovingToVpc | RestoringToClassic).

describeMovingAddresses_publicIps :: Lens' DescribeMovingAddresses (Maybe [Text])

One or more Elastic IP addresses.

describeMovingAddresses_nextToken :: Lens' DescribeMovingAddresses (Maybe Text)

The token for the next page of results.

describeMovingAddresses_dryRun :: Lens' DescribeMovingAddresses (Maybe Bool)

Checks whether you have the required permissions for the action, without actually making the request, and provides an error response. If you have the required permissions, the error response is DryRunOperation. Otherwise, it is UnauthorizedOperation.

describeMovingAddresses_maxResults :: Lens' DescribeMovingAddresses (Maybe Natural)

The maximum number of results to return for the request in a single page. The remaining results of the initial request can be seen by sending another request with the returned NextToken value. This value can be between 5 and 1000; if MaxResults is given a value outside of this range, an error is returned.

Default: If no value is provided, the default is 1000.

Destructuring the Response

data DescribeMovingAddressesResponse

See: newDescribeMovingAddressesResponse smart constructor.

Constructors

DescribeMovingAddressesResponse'

Fields

movingAddressStatuses :: Maybe [MovingAddressStatus]

The status for each Elastic IP address.

nextToken :: Maybe Text

The token to use to retrieve the next page of results. This value is null when there are no more results to return.

httpStatus :: Int

The response's http status code.

newDescribeMovingAddressesResponse

Arguments

:: Int
$sel:httpStatus:DescribeMovingAddressesResponse'

-> DescribeMovingAddressesResponse

Create a value of DescribeMovingAddressesResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:movingAddressStatuses:DescribeMovingAddressesResponse', describeMovingAddressesResponse_movingAddressStatuses - The status for each Elastic IP address.

$sel:nextToken:DescribeMovingAddresses', describeMovingAddressesResponse_nextToken - The token to use to retrieve the next page of results. This value is null when there are no more results to return.

$sel:httpStatus:DescribeMovingAddressesResponse', describeMovingAddressesResponse_httpStatus - The response's http status code.

Response Lenses

describeMovingAddressesResponse_movingAddressStatuses :: Lens' DescribeMovingAddressesResponse (Maybe [MovingAddressStatus])

The status for each Elastic IP address.

describeMovingAddressesResponse_nextToken :: Lens' DescribeMovingAddressesResponse (Maybe Text)

The token to use to retrieve the next page of results. This value is null when there are no more results to return.

describeMovingAddressesResponse_httpStatus :: Lens' DescribeMovingAddressesResponse Int

The response's http status code.