Amazonka.EC2.DescribePlacementGroups

Description

Describes the specified placement groups or all of your placement groups. For more information, see Placement groups in the Amazon EC2 User Guide.

Creating a Request

data DescribePlacementGroups

See: newDescribePlacementGroups smart constructor.

Constructors

DescribePlacementGroups'

Fields

filters :: Maybe [Filter]

The filters.

group-name - The name of the placement group.

state - The state of the placement group (pending | available | deleting | deleted).

strategy - The strategy of the placement group (cluster | spread | partition).

tag:<key> - The key/value combination of a tag assigned to the resource. Use the tag key in the filter name and the tag value as the filter value. For example, to find all resources that have a tag with the key Owner and the value TeamA, specify tag:Owner for the filter name and TeamA for the filter value.

tag-key - The key of a tag assigned to the resource. Use this filter to find all resources that have a tag with a specific key, regardless of the tag value.

groupNames :: Maybe [Text]

The names of the placement groups.

Default: Describes all your placement groups, or only those otherwise specified.

groupIds :: Maybe [Text]

The IDs of the placement groups.

dryRun :: Maybe Bool

Checks whether you have the required permissions for the action, without actually making the request, and provides an error response. If you have the required permissions, the error response is DryRunOperation. Otherwise, it is UnauthorizedOperation.

newDescribePlacementGroups :: DescribePlacementGroups

Create a value of DescribePlacementGroups with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:filters:DescribePlacementGroups', describePlacementGroups_filters - The filters.

group-name - The name of the placement group.

state - The state of the placement group (pending | available | deleting | deleted).

strategy - The strategy of the placement group (cluster | spread | partition).

tag:<key> - The key/value combination of a tag assigned to the resource. Use the tag key in the filter name and the tag value as the filter value. For example, to find all resources that have a tag with the key Owner and the value TeamA, specify tag:Owner for the filter name and TeamA for the filter value.

tag-key - The key of a tag assigned to the resource. Use this filter to find all resources that have a tag with a specific key, regardless of the tag value.

$sel:groupNames:DescribePlacementGroups', describePlacementGroups_groupNames - The names of the placement groups.

Default: Describes all your placement groups, or only those otherwise specified.

$sel:groupIds:DescribePlacementGroups', describePlacementGroups_groupIds - The IDs of the placement groups.

$sel:dryRun:DescribePlacementGroups', describePlacementGroups_dryRun - Checks whether you have the required permissions for the action, without actually making the request, and provides an error response. If you have the required permissions, the error response is DryRunOperation. Otherwise, it is UnauthorizedOperation.

Request Lenses

describePlacementGroups_filters :: Lens' DescribePlacementGroups (Maybe [Filter])

The filters.

group-name - The name of the placement group.

state - The state of the placement group (pending | available | deleting | deleted).

strategy - The strategy of the placement group (cluster | spread | partition).

tag:<key> - The key/value combination of a tag assigned to the resource. Use the tag key in the filter name and the tag value as the filter value. For example, to find all resources that have a tag with the key Owner and the value TeamA, specify tag:Owner for the filter name and TeamA for the filter value.

tag-key - The key of a tag assigned to the resource. Use this filter to find all resources that have a tag with a specific key, regardless of the tag value.

describePlacementGroups_groupNames :: Lens' DescribePlacementGroups (Maybe [Text])

The names of the placement groups.

Default: Describes all your placement groups, or only those otherwise specified.

describePlacementGroups_groupIds :: Lens' DescribePlacementGroups (Maybe [Text])

The IDs of the placement groups.

describePlacementGroups_dryRun :: Lens' DescribePlacementGroups (Maybe Bool)

Checks whether you have the required permissions for the action, without actually making the request, and provides an error response. If you have the required permissions, the error response is DryRunOperation. Otherwise, it is UnauthorizedOperation.

Destructuring the Response

data DescribePlacementGroupsResponse

See: newDescribePlacementGroupsResponse smart constructor.

Constructors

DescribePlacementGroupsResponse'

Fields

placementGroups :: Maybe [PlacementGroup]

Information about the placement groups.

httpStatus :: Int

The response's http status code.

newDescribePlacementGroupsResponse

Arguments

:: Int
$sel:httpStatus:DescribePlacementGroupsResponse'

-> DescribePlacementGroupsResponse

Create a value of DescribePlacementGroupsResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:placementGroups:DescribePlacementGroupsResponse', describePlacementGroupsResponse_placementGroups - Information about the placement groups.

$sel:httpStatus:DescribePlacementGroupsResponse', describePlacementGroupsResponse_httpStatus - The response's http status code.

Response Lenses

describePlacementGroupsResponse_placementGroups :: Lens' DescribePlacementGroupsResponse (Maybe [PlacementGroup])

Information about the placement groups.

describePlacementGroupsResponse_httpStatus :: Lens' DescribePlacementGroupsResponse Int

The response's http status code.