Amazonka.EC2.DescribeRouteTables

Description

Describes one or more of your route tables.

Each subnet in your VPC must be associated with a route table. If a subnet is not explicitly associated with any route table, it is implicitly associated with the main route table. This command does not return the subnet ID for implicit associations.

For more information, see Route tables in the Amazon Virtual Private Cloud User Guide.

This operation returns paginated results.

Creating a Request

data DescribeRouteTables

See: newDescribeRouteTables smart constructor.

Constructors

DescribeRouteTables'

Fields

filters :: Maybe [Filter]

One or more filters.

association.route-table-association-id - The ID of an association ID for the route table.

association.route-table-id - The ID of the route table involved in the association.

association.subnet-id - The ID of the subnet involved in the association.

association.main - Indicates whether the route table is the main route table for the VPC (true | false). Route tables that do not have an association ID are not returned in the response.

owner-id - The ID of the Amazon Web Services account that owns the route table.

route-table-id - The ID of the route table.

route.destination-cidr-block - The IPv4 CIDR range specified in a route in the table.

route.destination-ipv6-cidr-block - The IPv6 CIDR range specified in a route in the route table.

route.destination-prefix-list-id - The ID (prefix) of the Amazon Web Service specified in a route in the table.

route.egress-only-internet-gateway-id - The ID of an egress-only Internet gateway specified in a route in the route table.

route.gateway-id - The ID of a gateway specified in a route in the table.

route.instance-id - The ID of an instance specified in a route in the table.

route.nat-gateway-id - The ID of a NAT gateway.

route.transit-gateway-id - The ID of a transit gateway.

route.origin - Describes how the route was created. CreateRouteTable indicates that the route was automatically created when the route table was created; CreateRoute indicates that the route was manually added to the route table; EnableVgwRoutePropagation indicates that the route was propagated by route propagation.

route.state - The state of a route in the route table (active | blackhole). The blackhole state indicates that the route's target isn't available (for example, the specified gateway isn't attached to the VPC, the specified NAT instance has been terminated, and so on).

route.vpc-peering-connection-id - The ID of a VPC peering connection specified in a route in the table.

tag:<key> - The key/value combination of a tag assigned to the resource. Use the tag key in the filter name and the tag value as the filter value. For example, to find all resources that have a tag with the key Owner and the value TeamA, specify tag:Owner for the filter name and TeamA for the filter value.

tag-key - The key of a tag assigned to the resource. Use this filter to find all resources assigned a tag with a specific key, regardless of the tag value.

vpc-id - The ID of the VPC for the route table.

nextToken :: Maybe Text

The token for the next page of results.

dryRun :: Maybe Bool

Checks whether you have the required permissions for the action, without actually making the request, and provides an error response. If you have the required permissions, the error response is DryRunOperation. Otherwise, it is UnauthorizedOperation.

maxResults :: Maybe Natural

The maximum number of results to return with a single call. To retrieve the remaining results, make another call with the returned nextToken value.

routeTableIds :: Maybe [Text]

One or more route table IDs.

Default: Describes all your route tables.

newDescribeRouteTables :: DescribeRouteTables

Create a value of DescribeRouteTables with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:filters:DescribeRouteTables', describeRouteTables_filters - One or more filters.

association.route-table-association-id - The ID of an association ID for the route table.

association.route-table-id - The ID of the route table involved in the association.

association.subnet-id - The ID of the subnet involved in the association.

association.main - Indicates whether the route table is the main route table for the VPC (true | false). Route tables that do not have an association ID are not returned in the response.

owner-id - The ID of the Amazon Web Services account that owns the route table.

route-table-id - The ID of the route table.

route.destination-cidr-block - The IPv4 CIDR range specified in a route in the table.

route.destination-ipv6-cidr-block - The IPv6 CIDR range specified in a route in the route table.

route.destination-prefix-list-id - The ID (prefix) of the Amazon Web Service specified in a route in the table.

route.egress-only-internet-gateway-id - The ID of an egress-only Internet gateway specified in a route in the route table.

route.gateway-id - The ID of a gateway specified in a route in the table.

route.instance-id - The ID of an instance specified in a route in the table.

route.nat-gateway-id - The ID of a NAT gateway.

route.transit-gateway-id - The ID of a transit gateway.

route.origin - Describes how the route was created. CreateRouteTable indicates that the route was automatically created when the route table was created; CreateRoute indicates that the route was manually added to the route table; EnableVgwRoutePropagation indicates that the route was propagated by route propagation.

route.state - The state of a route in the route table (active | blackhole). The blackhole state indicates that the route's target isn't available (for example, the specified gateway isn't attached to the VPC, the specified NAT instance has been terminated, and so on).

route.vpc-peering-connection-id - The ID of a VPC peering connection specified in a route in the table.

tag:<key> - The key/value combination of a tag assigned to the resource. Use the tag key in the filter name and the tag value as the filter value. For example, to find all resources that have a tag with the key Owner and the value TeamA, specify tag:Owner for the filter name and TeamA for the filter value.

tag-key - The key of a tag assigned to the resource. Use this filter to find all resources assigned a tag with a specific key, regardless of the tag value.

vpc-id - The ID of the VPC for the route table.

$sel:nextToken:DescribeRouteTables', describeRouteTables_nextToken - The token for the next page of results.

$sel:dryRun:DescribeRouteTables', describeRouteTables_dryRun - Checks whether you have the required permissions for the action, without actually making the request, and provides an error response. If you have the required permissions, the error response is DryRunOperation. Otherwise, it is UnauthorizedOperation.

$sel:maxResults:DescribeRouteTables', describeRouteTables_maxResults - The maximum number of results to return with a single call. To retrieve the remaining results, make another call with the returned nextToken value.

$sel:routeTableIds:DescribeRouteTables', describeRouteTables_routeTableIds - One or more route table IDs.

Default: Describes all your route tables.

Request Lenses

describeRouteTables_filters :: Lens' DescribeRouteTables (Maybe [Filter])

One or more filters.

association.route-table-association-id - The ID of an association ID for the route table.

association.route-table-id - The ID of the route table involved in the association.

association.subnet-id - The ID of the subnet involved in the association.

association.main - Indicates whether the route table is the main route table for the VPC (true | false). Route tables that do not have an association ID are not returned in the response.

owner-id - The ID of the Amazon Web Services account that owns the route table.

route-table-id - The ID of the route table.

route.destination-cidr-block - The IPv4 CIDR range specified in a route in the table.

route.destination-ipv6-cidr-block - The IPv6 CIDR range specified in a route in the route table.

route.destination-prefix-list-id - The ID (prefix) of the Amazon Web Service specified in a route in the table.

route.egress-only-internet-gateway-id - The ID of an egress-only Internet gateway specified in a route in the route table.

route.gateway-id - The ID of a gateway specified in a route in the table.

route.instance-id - The ID of an instance specified in a route in the table.

route.nat-gateway-id - The ID of a NAT gateway.

route.transit-gateway-id - The ID of a transit gateway.

route.origin - Describes how the route was created. CreateRouteTable indicates that the route was automatically created when the route table was created; CreateRoute indicates that the route was manually added to the route table; EnableVgwRoutePropagation indicates that the route was propagated by route propagation.

route.state - The state of a route in the route table (active | blackhole). The blackhole state indicates that the route's target isn't available (for example, the specified gateway isn't attached to the VPC, the specified NAT instance has been terminated, and so on).

route.vpc-peering-connection-id - The ID of a VPC peering connection specified in a route in the table.

tag:<key> - The key/value combination of a tag assigned to the resource. Use the tag key in the filter name and the tag value as the filter value. For example, to find all resources that have a tag with the key Owner and the value TeamA, specify tag:Owner for the filter name and TeamA for the filter value.

tag-key - The key of a tag assigned to the resource. Use this filter to find all resources assigned a tag with a specific key, regardless of the tag value.

vpc-id - The ID of the VPC for the route table.

describeRouteTables_nextToken :: Lens' DescribeRouteTables (Maybe Text)

The token for the next page of results.

describeRouteTables_dryRun :: Lens' DescribeRouteTables (Maybe Bool)

Checks whether you have the required permissions for the action, without actually making the request, and provides an error response. If you have the required permissions, the error response is DryRunOperation. Otherwise, it is UnauthorizedOperation.

describeRouteTables_maxResults :: Lens' DescribeRouteTables (Maybe Natural)

The maximum number of results to return with a single call. To retrieve the remaining results, make another call with the returned nextToken value.

describeRouteTables_routeTableIds :: Lens' DescribeRouteTables (Maybe [Text])

One or more route table IDs.

Default: Describes all your route tables.

Destructuring the Response

data DescribeRouteTablesResponse

Contains the output of DescribeRouteTables.

See: newDescribeRouteTablesResponse smart constructor.

Constructors

DescribeRouteTablesResponse'

Fields

nextToken :: Maybe Text

The token to use to retrieve the next page of results. This value is null when there are no more results to return.

routeTables :: Maybe [RouteTable]

Information about one or more route tables.

httpStatus :: Int

The response's http status code.

newDescribeRouteTablesResponse

Arguments

:: Int
$sel:httpStatus:DescribeRouteTablesResponse'

-> DescribeRouteTablesResponse

Create a value of DescribeRouteTablesResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:nextToken:DescribeRouteTables', describeRouteTablesResponse_nextToken - The token to use to retrieve the next page of results. This value is null when there are no more results to return.

$sel:routeTables:DescribeRouteTablesResponse', describeRouteTablesResponse_routeTables - Information about one or more route tables.

$sel:httpStatus:DescribeRouteTablesResponse', describeRouteTablesResponse_httpStatus - The response's http status code.

Response Lenses

describeRouteTablesResponse_nextToken :: Lens' DescribeRouteTablesResponse (Maybe Text)

The token to use to retrieve the next page of results. This value is null when there are no more results to return.

describeRouteTablesResponse_routeTables :: Lens' DescribeRouteTablesResponse (Maybe [RouteTable])

Information about one or more route tables.

describeRouteTablesResponse_httpStatus :: Lens' DescribeRouteTablesResponse Int

The response's http status code.