Amazonka.EC2.DescribeScheduledInstances

Description

Describes the specified Scheduled Instances or all your Scheduled Instances.

This operation returns paginated results.

Creating a Request

data DescribeScheduledInstances

Contains the parameters for DescribeScheduledInstances.

See: newDescribeScheduledInstances smart constructor.

Constructors

DescribeScheduledInstances'

Fields

filters :: Maybe [Filter]

The filters.

availability-zone - The Availability Zone (for example, us-west-2a).

instance-type - The instance type (for example, c4.large).

network-platform - The network platform (EC2-Classic or EC2-VPC).

platform - The platform (Linux/UNIX or Windows).

slotStartTimeRange :: Maybe SlotStartTimeRangeRequest

The time period for the first schedule to start.

nextToken :: Maybe Text

The token for the next set of results.

scheduledInstanceIds :: Maybe [Text]

The Scheduled Instance IDs.

dryRun :: Maybe Bool

Checks whether you have the required permissions for the action, without actually making the request, and provides an error response. If you have the required permissions, the error response is DryRunOperation. Otherwise, it is UnauthorizedOperation.

maxResults :: Maybe Int

The maximum number of results to return in a single call. This value can be between 5 and 300. The default value is 100. To retrieve the remaining results, make another call with the returned NextToken value.

newDescribeScheduledInstances :: DescribeScheduledInstances

Create a value of DescribeScheduledInstances with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:filters:DescribeScheduledInstances', describeScheduledInstances_filters - The filters.

availability-zone - The Availability Zone (for example, us-west-2a).

instance-type - The instance type (for example, c4.large).

network-platform - The network platform (EC2-Classic or EC2-VPC).

platform - The platform (Linux/UNIX or Windows).

$sel:slotStartTimeRange:DescribeScheduledInstances', describeScheduledInstances_slotStartTimeRange - The time period for the first schedule to start.

$sel:nextToken:DescribeScheduledInstances', describeScheduledInstances_nextToken - The token for the next set of results.

$sel:scheduledInstanceIds:DescribeScheduledInstances', describeScheduledInstances_scheduledInstanceIds - The Scheduled Instance IDs.

$sel:dryRun:DescribeScheduledInstances', describeScheduledInstances_dryRun - Checks whether you have the required permissions for the action, without actually making the request, and provides an error response. If you have the required permissions, the error response is DryRunOperation. Otherwise, it is UnauthorizedOperation.

$sel:maxResults:DescribeScheduledInstances', describeScheduledInstances_maxResults - The maximum number of results to return in a single call. This value can be between 5 and 300. The default value is 100. To retrieve the remaining results, make another call with the returned NextToken value.

Request Lenses

describeScheduledInstances_filters :: Lens' DescribeScheduledInstances (Maybe [Filter])

The filters.

availability-zone - The Availability Zone (for example, us-west-2a).

instance-type - The instance type (for example, c4.large).

network-platform - The network platform (EC2-Classic or EC2-VPC).

platform - The platform (Linux/UNIX or Windows).

describeScheduledInstances_slotStartTimeRange :: Lens' DescribeScheduledInstances (Maybe SlotStartTimeRangeRequest)

The time period for the first schedule to start.

describeScheduledInstances_nextToken :: Lens' DescribeScheduledInstances (Maybe Text)

The token for the next set of results.

describeScheduledInstances_scheduledInstanceIds :: Lens' DescribeScheduledInstances (Maybe [Text])

The Scheduled Instance IDs.

describeScheduledInstances_dryRun :: Lens' DescribeScheduledInstances (Maybe Bool)

Checks whether you have the required permissions for the action, without actually making the request, and provides an error response. If you have the required permissions, the error response is DryRunOperation. Otherwise, it is UnauthorizedOperation.

describeScheduledInstances_maxResults :: Lens' DescribeScheduledInstances (Maybe Int)

The maximum number of results to return in a single call. This value can be between 5 and 300. The default value is 100. To retrieve the remaining results, make another call with the returned NextToken value.

Destructuring the Response

data DescribeScheduledInstancesResponse

Contains the output of DescribeScheduledInstances.

See: newDescribeScheduledInstancesResponse smart constructor.

Constructors

DescribeScheduledInstancesResponse'

Fields

nextToken :: Maybe Text

The token required to retrieve the next set of results. This value is null when there are no more results to return.

scheduledInstanceSet :: Maybe [ScheduledInstance]

Information about the Scheduled Instances.

httpStatus :: Int

The response's http status code.

newDescribeScheduledInstancesResponse

Arguments

:: Int
$sel:httpStatus:DescribeScheduledInstancesResponse'

-> DescribeScheduledInstancesResponse

Create a value of DescribeScheduledInstancesResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:nextToken:DescribeScheduledInstances', describeScheduledInstancesResponse_nextToken - The token required to retrieve the next set of results. This value is null when there are no more results to return.

$sel:scheduledInstanceSet:DescribeScheduledInstancesResponse', describeScheduledInstancesResponse_scheduledInstanceSet - Information about the Scheduled Instances.

$sel:httpStatus:DescribeScheduledInstancesResponse', describeScheduledInstancesResponse_httpStatus - The response's http status code.

Response Lenses

describeScheduledInstancesResponse_nextToken :: Lens' DescribeScheduledInstancesResponse (Maybe Text)

The token required to retrieve the next set of results. This value is null when there are no more results to return.

describeScheduledInstancesResponse_scheduledInstanceSet :: Lens' DescribeScheduledInstancesResponse (Maybe [ScheduledInstance])

Information about the Scheduled Instances.

describeScheduledInstancesResponse_httpStatus :: Lens' DescribeScheduledInstancesResponse Int

The response's http status code.