Amazonka.EC2.DescribeSpotFleetInstances

Description

Describes the running instances for the specified Spot Fleet.

This operation returns paginated results.

Creating a Request

data DescribeSpotFleetInstances

Contains the parameters for DescribeSpotFleetInstances.

See: newDescribeSpotFleetInstances smart constructor.

Constructors

DescribeSpotFleetInstances'

Fields

nextToken :: Maybe Text

The token for the next set of results.

dryRun :: Maybe Bool

Checks whether you have the required permissions for the action, without actually making the request, and provides an error response. If you have the required permissions, the error response is DryRunOperation. Otherwise, it is UnauthorizedOperation.

maxResults :: Maybe Natural

The maximum number of results to return in a single call. Specify a value between 1 and 1000. The default value is 1000. To retrieve the remaining results, make another call with the returned NextToken value.

spotFleetRequestId :: Text

The ID of the Spot Fleet request.

newDescribeSpotFleetInstances

Arguments

:: Text
$sel:spotFleetRequestId:DescribeSpotFleetInstances'

-> DescribeSpotFleetInstances

Create a value of DescribeSpotFleetInstances with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:nextToken:DescribeSpotFleetInstances', describeSpotFleetInstances_nextToken - The token for the next set of results.

$sel:dryRun:DescribeSpotFleetInstances', describeSpotFleetInstances_dryRun - Checks whether you have the required permissions for the action, without actually making the request, and provides an error response. If you have the required permissions, the error response is DryRunOperation. Otherwise, it is UnauthorizedOperation.

$sel:maxResults:DescribeSpotFleetInstances', describeSpotFleetInstances_maxResults - The maximum number of results to return in a single call. Specify a value between 1 and 1000. The default value is 1000. To retrieve the remaining results, make another call with the returned NextToken value.

$sel:spotFleetRequestId:DescribeSpotFleetInstances', describeSpotFleetInstances_spotFleetRequestId - The ID of the Spot Fleet request.

Request Lenses

describeSpotFleetInstances_nextToken :: Lens' DescribeSpotFleetInstances (Maybe Text)

The token for the next set of results.

describeSpotFleetInstances_dryRun :: Lens' DescribeSpotFleetInstances (Maybe Bool)

Checks whether you have the required permissions for the action, without actually making the request, and provides an error response. If you have the required permissions, the error response is DryRunOperation. Otherwise, it is UnauthorizedOperation.

describeSpotFleetInstances_maxResults :: Lens' DescribeSpotFleetInstances (Maybe Natural)

The maximum number of results to return in a single call. Specify a value between 1 and 1000. The default value is 1000. To retrieve the remaining results, make another call with the returned NextToken value.

describeSpotFleetInstances_spotFleetRequestId :: Lens' DescribeSpotFleetInstances Text

The ID of the Spot Fleet request.

Destructuring the Response

data DescribeSpotFleetInstancesResponse

Contains the output of DescribeSpotFleetInstances.

See: newDescribeSpotFleetInstancesResponse smart constructor.

Constructors

DescribeSpotFleetInstancesResponse'

Fields

nextToken :: Maybe Text

The token required to retrieve the next set of results. This value is null when there are no more results to return.

spotFleetRequestId :: Maybe Text

The ID of the Spot Fleet request.

activeInstances :: Maybe [ActiveInstance]

The running instances. This list is refreshed periodically and might be out of date.

httpStatus :: Int

The response's http status code.

newDescribeSpotFleetInstancesResponse

Arguments

:: Int
$sel:httpStatus:DescribeSpotFleetInstancesResponse'

-> DescribeSpotFleetInstancesResponse

Create a value of DescribeSpotFleetInstancesResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:nextToken:DescribeSpotFleetInstances', describeSpotFleetInstancesResponse_nextToken - The token required to retrieve the next set of results. This value is null when there are no more results to return.

$sel:spotFleetRequestId:DescribeSpotFleetInstances', describeSpotFleetInstancesResponse_spotFleetRequestId - The ID of the Spot Fleet request.

$sel:activeInstances:DescribeSpotFleetInstancesResponse', describeSpotFleetInstancesResponse_activeInstances - The running instances. This list is refreshed periodically and might be out of date.

$sel:httpStatus:DescribeSpotFleetInstancesResponse', describeSpotFleetInstancesResponse_httpStatus - The response's http status code.

Response Lenses

describeSpotFleetInstancesResponse_nextToken :: Lens' DescribeSpotFleetInstancesResponse (Maybe Text)

The token required to retrieve the next set of results. This value is null when there are no more results to return.

describeSpotFleetInstancesResponse_spotFleetRequestId :: Lens' DescribeSpotFleetInstancesResponse (Maybe Text)

The ID of the Spot Fleet request.

describeSpotFleetInstancesResponse_activeInstances :: Lens' DescribeSpotFleetInstancesResponse (Maybe [ActiveInstance])

The running instances. This list is refreshed periodically and might be out of date.

describeSpotFleetInstancesResponse_httpStatus :: Lens' DescribeSpotFleetInstancesResponse Int

The response's http status code.