Amazonka.EC2.DescribeTransitGatewayAttachments

Description

Describes one or more attachments between resources and transit gateways. By default, all attachments are described. Alternatively, you can filter the results by attachment ID, attachment state, resource ID, or resource owner.

This operation returns paginated results.

Creating a Request

data DescribeTransitGatewayAttachments

See: newDescribeTransitGatewayAttachments smart constructor.

Constructors

DescribeTransitGatewayAttachments'

Fields

filters :: Maybe [Filter]

One or more filters. The possible values are:

association.state - The state of the association (associating | associated | disassociating).

association.transit-gateway-route-table-id - The ID of the route table for the transit gateway.

resource-id - The ID of the resource.

resource-owner-id - The ID of the Amazon Web Services account that owns the resource.

resource-type - The resource type. Valid values are vpc | vpn | direct-connect-gateway | peering | connect.

state - The state of the attachment. Valid values are available | deleted | deleting | failed | failing | initiatingRequest | modifying | pendingAcceptance | pending | rollingBack | rejected | rejecting.

transit-gateway-attachment-id - The ID of the attachment.

transit-gateway-id - The ID of the transit gateway.

transit-gateway-owner-id - The ID of the Amazon Web Services account that owns the transit gateway.

nextToken :: Maybe Text

The token for the next page of results.

transitGatewayAttachmentIds :: Maybe [Text]

The IDs of the attachments.

dryRun :: Maybe Bool

Checks whether you have the required permissions for the action, without actually making the request, and provides an error response. If you have the required permissions, the error response is DryRunOperation. Otherwise, it is UnauthorizedOperation.

maxResults :: Maybe Natural

The maximum number of results to return with a single call. To retrieve the remaining results, make another call with the returned nextToken value.

newDescribeTransitGatewayAttachments :: DescribeTransitGatewayAttachments

Create a value of DescribeTransitGatewayAttachments with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:filters:DescribeTransitGatewayAttachments', describeTransitGatewayAttachments_filters - One or more filters. The possible values are:

association.state - The state of the association (associating | associated | disassociating).

association.transit-gateway-route-table-id - The ID of the route table for the transit gateway.

resource-id - The ID of the resource.

resource-owner-id - The ID of the Amazon Web Services account that owns the resource.

resource-type - The resource type. Valid values are vpc | vpn | direct-connect-gateway | peering | connect.

state - The state of the attachment. Valid values are available | deleted | deleting | failed | failing | initiatingRequest | modifying | pendingAcceptance | pending | rollingBack | rejected | rejecting.

transit-gateway-attachment-id - The ID of the attachment.

transit-gateway-id - The ID of the transit gateway.

transit-gateway-owner-id - The ID of the Amazon Web Services account that owns the transit gateway.

$sel:nextToken:DescribeTransitGatewayAttachments', describeTransitGatewayAttachments_nextToken - The token for the next page of results.

$sel:transitGatewayAttachmentIds:DescribeTransitGatewayAttachments', describeTransitGatewayAttachments_transitGatewayAttachmentIds - The IDs of the attachments.

$sel:dryRun:DescribeTransitGatewayAttachments', describeTransitGatewayAttachments_dryRun - Checks whether you have the required permissions for the action, without actually making the request, and provides an error response. If you have the required permissions, the error response is DryRunOperation. Otherwise, it is UnauthorizedOperation.

$sel:maxResults:DescribeTransitGatewayAttachments', describeTransitGatewayAttachments_maxResults - The maximum number of results to return with a single call. To retrieve the remaining results, make another call with the returned nextToken value.

Request Lenses

describeTransitGatewayAttachments_filters :: Lens' DescribeTransitGatewayAttachments (Maybe [Filter])

One or more filters. The possible values are:

association.state - The state of the association (associating | associated | disassociating).

association.transit-gateway-route-table-id - The ID of the route table for the transit gateway.

resource-id - The ID of the resource.

resource-owner-id - The ID of the Amazon Web Services account that owns the resource.

resource-type - The resource type. Valid values are vpc | vpn | direct-connect-gateway | peering | connect.

state - The state of the attachment. Valid values are available | deleted | deleting | failed | failing | initiatingRequest | modifying | pendingAcceptance | pending | rollingBack | rejected | rejecting.

transit-gateway-attachment-id - The ID of the attachment.

transit-gateway-id - The ID of the transit gateway.

transit-gateway-owner-id - The ID of the Amazon Web Services account that owns the transit gateway.

describeTransitGatewayAttachments_nextToken :: Lens' DescribeTransitGatewayAttachments (Maybe Text)

The token for the next page of results.

describeTransitGatewayAttachments_transitGatewayAttachmentIds :: Lens' DescribeTransitGatewayAttachments (Maybe [Text])

The IDs of the attachments.

describeTransitGatewayAttachments_dryRun :: Lens' DescribeTransitGatewayAttachments (Maybe Bool)

Checks whether you have the required permissions for the action, without actually making the request, and provides an error response. If you have the required permissions, the error response is DryRunOperation. Otherwise, it is UnauthorizedOperation.

describeTransitGatewayAttachments_maxResults :: Lens' DescribeTransitGatewayAttachments (Maybe Natural)

The maximum number of results to return with a single call. To retrieve the remaining results, make another call with the returned nextToken value.

Destructuring the Response

data DescribeTransitGatewayAttachmentsResponse

See: newDescribeTransitGatewayAttachmentsResponse smart constructor.

Constructors

DescribeTransitGatewayAttachmentsResponse'

Fields

nextToken :: Maybe Text

The token to use to retrieve the next page of results. This value is null when there are no more results to return.

transitGatewayAttachments :: Maybe [TransitGatewayAttachment]

Information about the attachments.

httpStatus :: Int

The response's http status code.

newDescribeTransitGatewayAttachmentsResponse

Arguments

:: Int
$sel:httpStatus:DescribeTransitGatewayAttachmentsResponse'

-> DescribeTransitGatewayAttachmentsResponse

Create a value of DescribeTransitGatewayAttachmentsResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:nextToken:DescribeTransitGatewayAttachments', describeTransitGatewayAttachmentsResponse_nextToken - The token to use to retrieve the next page of results. This value is null when there are no more results to return.

$sel:transitGatewayAttachments:DescribeTransitGatewayAttachmentsResponse', describeTransitGatewayAttachmentsResponse_transitGatewayAttachments - Information about the attachments.

$sel:httpStatus:DescribeTransitGatewayAttachmentsResponse', describeTransitGatewayAttachmentsResponse_httpStatus - The response's http status code.

Response Lenses

describeTransitGatewayAttachmentsResponse_nextToken :: Lens' DescribeTransitGatewayAttachmentsResponse (Maybe Text)

The token to use to retrieve the next page of results. This value is null when there are no more results to return.

describeTransitGatewayAttachmentsResponse_transitGatewayAttachments :: Lens' DescribeTransitGatewayAttachmentsResponse (Maybe [TransitGatewayAttachment])

Information about the attachments.

describeTransitGatewayAttachmentsResponse_httpStatus :: Lens' DescribeTransitGatewayAttachmentsResponse Int

The response's http status code.