Amazonka.EC2.DescribeVolumes

Description

Describes the specified EBS volumes or all of your EBS volumes.

If you are describing a long list of volumes, we recommend that you paginate the output to make the list more manageable. The MaxResults parameter sets the maximum number of results returned in a single page. If the list of results exceeds your MaxResults value, then that number of results is returned along with a NextToken value that can be passed to a subsequent DescribeVolumes request to retrieve the remaining results.

For more information about EBS volumes, see Amazon EBS volumes in the Amazon Elastic Compute Cloud User Guide.

This operation returns paginated results.

Creating a Request

data DescribeVolumes

See: newDescribeVolumes smart constructor.

Constructors

DescribeVolumes'

Fields

filters :: Maybe [Filter]

The filters.

attachment.attach-time - The time stamp when the attachment initiated.

attachment.delete-on-termination - Whether the volume is deleted on instance termination.

attachment.device - The device name specified in the block device mapping (for example, /dev/sda1).

attachment.instance-id - The ID of the instance the volume is attached to.

attachment.status - The attachment state (attaching | attached | detaching).

availability-zone - The Availability Zone in which the volume was created.

create-time - The time stamp when the volume was created.

encrypted - Indicates whether the volume is encrypted (true | false)

multi-attach-enabled - Indicates whether the volume is enabled for Multi-Attach (true | false)

fast-restored - Indicates whether the volume was created from a snapshot that is enabled for fast snapshot restore (true | false).

size - The size of the volume, in GiB.

snapshot-id - The snapshot from which the volume was created.

status - The state of the volume (creating | available | in-use | deleting | deleted | error).

tag:<key> - The key/value combination of a tag assigned to the resource. Use the tag key in the filter name and the tag value as the filter value. For example, to find all resources that have a tag with the key Owner and the value TeamA, specify tag:Owner for the filter name and TeamA for the filter value.

tag-key - The key of a tag assigned to the resource. Use this filter to find all resources assigned a tag with a specific key, regardless of the tag value.

volume-id - The volume ID.

volume-type - The Amazon EBS volume type (gp2 | gp3 | io1 | io2 | st1 | sc1| standard)

volumeIds :: Maybe [Text]

The volume IDs.

nextToken :: Maybe Text

The NextToken value returned from a previous paginated DescribeVolumes request where MaxResults was used and the results exceeded the value of that parameter. Pagination continues from the end of the previous results that returned the NextToken value. This value is null when there are no more results to return.

dryRun :: Maybe Bool

Checks whether you have the required permissions for the action, without actually making the request, and provides an error response. If you have the required permissions, the error response is DryRunOperation. Otherwise, it is UnauthorizedOperation.

maxResults :: Maybe Int

The maximum number of volume results returned by DescribeVolumes in paginated output. When this parameter is used, DescribeVolumes only returns MaxResults results in a single page along with a NextToken response element. The remaining results of the initial request can be seen by sending another DescribeVolumes request with the returned NextToken value. This value can be between 5 and 500; if MaxResults is given a value larger than 500, only 500 results are returned. If this parameter is not used, then DescribeVolumes returns all results. You cannot specify this parameter and the volume IDs parameter in the same request.

newDescribeVolumes :: DescribeVolumes

Create a value of DescribeVolumes with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:filters:DescribeVolumes', describeVolumes_filters - The filters.

attachment.attach-time - The time stamp when the attachment initiated.

attachment.delete-on-termination - Whether the volume is deleted on instance termination.

attachment.device - The device name specified in the block device mapping (for example, /dev/sda1).

attachment.instance-id - The ID of the instance the volume is attached to.

attachment.status - The attachment state (attaching | attached | detaching).

availability-zone - The Availability Zone in which the volume was created.

create-time - The time stamp when the volume was created.

encrypted - Indicates whether the volume is encrypted (true | false)

multi-attach-enabled - Indicates whether the volume is enabled for Multi-Attach (true | false)

fast-restored - Indicates whether the volume was created from a snapshot that is enabled for fast snapshot restore (true | false).

size - The size of the volume, in GiB.

snapshot-id - The snapshot from which the volume was created.

status - The state of the volume (creating | available | in-use | deleting | deleted | error).

tag:<key> - The key/value combination of a tag assigned to the resource. Use the tag key in the filter name and the tag value as the filter value. For example, to find all resources that have a tag with the key Owner and the value TeamA, specify tag:Owner for the filter name and TeamA for the filter value.

tag-key - The key of a tag assigned to the resource. Use this filter to find all resources assigned a tag with a specific key, regardless of the tag value.

volume-id - The volume ID.

volume-type - The Amazon EBS volume type (gp2 | gp3 | io1 | io2 | st1 | sc1| standard)

$sel:volumeIds:DescribeVolumes', describeVolumes_volumeIds - The volume IDs.

$sel:nextToken:DescribeVolumes', describeVolumes_nextToken - The NextToken value returned from a previous paginated DescribeVolumes request where MaxResults was used and the results exceeded the value of that parameter. Pagination continues from the end of the previous results that returned the NextToken value. This value is null when there are no more results to return.

$sel:dryRun:DescribeVolumes', describeVolumes_dryRun - Checks whether you have the required permissions for the action, without actually making the request, and provides an error response. If you have the required permissions, the error response is DryRunOperation. Otherwise, it is UnauthorizedOperation.

$sel:maxResults:DescribeVolumes', describeVolumes_maxResults - The maximum number of volume results returned by DescribeVolumes in paginated output. When this parameter is used, DescribeVolumes only returns MaxResults results in a single page along with a NextToken response element. The remaining results of the initial request can be seen by sending another DescribeVolumes request with the returned NextToken value. This value can be between 5 and 500; if MaxResults is given a value larger than 500, only 500 results are returned. If this parameter is not used, then DescribeVolumes returns all results. You cannot specify this parameter and the volume IDs parameter in the same request.

Request Lenses

describeVolumes_filters :: Lens' DescribeVolumes (Maybe [Filter])

The filters.

attachment.attach-time - The time stamp when the attachment initiated.

attachment.delete-on-termination - Whether the volume is deleted on instance termination.

attachment.device - The device name specified in the block device mapping (for example, /dev/sda1).

attachment.instance-id - The ID of the instance the volume is attached to.

attachment.status - The attachment state (attaching | attached | detaching).

availability-zone - The Availability Zone in which the volume was created.

create-time - The time stamp when the volume was created.

encrypted - Indicates whether the volume is encrypted (true | false)

multi-attach-enabled - Indicates whether the volume is enabled for Multi-Attach (true | false)

fast-restored - Indicates whether the volume was created from a snapshot that is enabled for fast snapshot restore (true | false).

size - The size of the volume, in GiB.

snapshot-id - The snapshot from which the volume was created.

status - The state of the volume (creating | available | in-use | deleting | deleted | error).

tag:<key> - The key/value combination of a tag assigned to the resource. Use the tag key in the filter name and the tag value as the filter value. For example, to find all resources that have a tag with the key Owner and the value TeamA, specify tag:Owner for the filter name and TeamA for the filter value.

tag-key - The key of a tag assigned to the resource. Use this filter to find all resources assigned a tag with a specific key, regardless of the tag value.

volume-id - The volume ID.

volume-type - The Amazon EBS volume type (gp2 | gp3 | io1 | io2 | st1 | sc1| standard)

describeVolumes_volumeIds :: Lens' DescribeVolumes (Maybe [Text])

The volume IDs.

describeVolumes_nextToken :: Lens' DescribeVolumes (Maybe Text)

The NextToken value returned from a previous paginated DescribeVolumes request where MaxResults was used and the results exceeded the value of that parameter. Pagination continues from the end of the previous results that returned the NextToken value. This value is null when there are no more results to return.

describeVolumes_dryRun :: Lens' DescribeVolumes (Maybe Bool)

Checks whether you have the required permissions for the action, without actually making the request, and provides an error response. If you have the required permissions, the error response is DryRunOperation. Otherwise, it is UnauthorizedOperation.

describeVolumes_maxResults :: Lens' DescribeVolumes (Maybe Int)

The maximum number of volume results returned by DescribeVolumes in paginated output. When this parameter is used, DescribeVolumes only returns MaxResults results in a single page along with a NextToken response element. The remaining results of the initial request can be seen by sending another DescribeVolumes request with the returned NextToken value. This value can be between 5 and 500; if MaxResults is given a value larger than 500, only 500 results are returned. If this parameter is not used, then DescribeVolumes returns all results. You cannot specify this parameter and the volume IDs parameter in the same request.

Destructuring the Response

data DescribeVolumesResponse

See: newDescribeVolumesResponse smart constructor.

Constructors

DescribeVolumesResponse'

Fields

nextToken :: Maybe Text

The NextToken value to include in a future DescribeVolumes request. When the results of a DescribeVolumes request exceed MaxResults, this value can be used to retrieve the next page of results. This value is null when there are no more results to return.

volumes :: Maybe [Volume]

Information about the volumes.

httpStatus :: Int

The response's http status code.

newDescribeVolumesResponse

Arguments

:: Int
$sel:httpStatus:DescribeVolumesResponse'

-> DescribeVolumesResponse

Create a value of DescribeVolumesResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:nextToken:DescribeVolumes', describeVolumesResponse_nextToken - The NextToken value to include in a future DescribeVolumes request. When the results of a DescribeVolumes request exceed MaxResults, this value can be used to retrieve the next page of results. This value is null when there are no more results to return.

$sel:volumes:DescribeVolumesResponse', describeVolumesResponse_volumes - Information about the volumes.

$sel:httpStatus:DescribeVolumesResponse', describeVolumesResponse_httpStatus - The response's http status code.

Response Lenses

describeVolumesResponse_nextToken :: Lens' DescribeVolumesResponse (Maybe Text)

The NextToken value to include in a future DescribeVolumes request. When the results of a DescribeVolumes request exceed MaxResults, this value can be used to retrieve the next page of results. This value is null when there are no more results to return.

describeVolumesResponse_volumes :: Lens' DescribeVolumesResponse (Maybe [Volume])

Information about the volumes.

describeVolumesResponse_httpStatus :: Lens' DescribeVolumesResponse Int

The response's http status code.