Amazonka.EC2.DescribeVpcClassicLink

Description

Describes the ClassicLink status of one or more VPCs.

Creating a Request

data DescribeVpcClassicLink

See: newDescribeVpcClassicLink smart constructor.

Constructors

DescribeVpcClassicLink'

Fields

filters :: Maybe [Filter]

One or more filters.

is-classic-link-enabled - Whether the VPC is enabled for ClassicLink (true | false).

tag:<key> - The key/value combination of a tag assigned to the resource. Use the tag key in the filter name and the tag value as the filter value. For example, to find all resources that have a tag with the key Owner and the value TeamA, specify tag:Owner for the filter name and TeamA for the filter value.

tag-key - The key of a tag assigned to the resource. Use this filter to find all resources assigned a tag with a specific key, regardless of the tag value.

vpcIds :: Maybe [Text]

One or more VPCs for which you want to describe the ClassicLink status.

dryRun :: Maybe Bool

Checks whether you have the required permissions for the action, without actually making the request, and provides an error response. If you have the required permissions, the error response is DryRunOperation. Otherwise, it is UnauthorizedOperation.

newDescribeVpcClassicLink :: DescribeVpcClassicLink

Create a value of DescribeVpcClassicLink with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:filters:DescribeVpcClassicLink', describeVpcClassicLink_filters - One or more filters.

is-classic-link-enabled - Whether the VPC is enabled for ClassicLink (true | false).

tag:<key> - The key/value combination of a tag assigned to the resource. Use the tag key in the filter name and the tag value as the filter value. For example, to find all resources that have a tag with the key Owner and the value TeamA, specify tag:Owner for the filter name and TeamA for the filter value.

tag-key - The key of a tag assigned to the resource. Use this filter to find all resources assigned a tag with a specific key, regardless of the tag value.

$sel:vpcIds:DescribeVpcClassicLink', describeVpcClassicLink_vpcIds - One or more VPCs for which you want to describe the ClassicLink status.

$sel:dryRun:DescribeVpcClassicLink', describeVpcClassicLink_dryRun - Checks whether you have the required permissions for the action, without actually making the request, and provides an error response. If you have the required permissions, the error response is DryRunOperation. Otherwise, it is UnauthorizedOperation.

Request Lenses

describeVpcClassicLink_filters :: Lens' DescribeVpcClassicLink (Maybe [Filter])

One or more filters.

is-classic-link-enabled - Whether the VPC is enabled for ClassicLink (true | false).

tag:<key> - The key/value combination of a tag assigned to the resource. Use the tag key in the filter name and the tag value as the filter value. For example, to find all resources that have a tag with the key Owner and the value TeamA, specify tag:Owner for the filter name and TeamA for the filter value.

tag-key - The key of a tag assigned to the resource. Use this filter to find all resources assigned a tag with a specific key, regardless of the tag value.

describeVpcClassicLink_vpcIds :: Lens' DescribeVpcClassicLink (Maybe [Text])

One or more VPCs for which you want to describe the ClassicLink status.

describeVpcClassicLink_dryRun :: Lens' DescribeVpcClassicLink (Maybe Bool)

Checks whether you have the required permissions for the action, without actually making the request, and provides an error response. If you have the required permissions, the error response is DryRunOperation. Otherwise, it is UnauthorizedOperation.

Destructuring the Response

data DescribeVpcClassicLinkResponse

See: newDescribeVpcClassicLinkResponse smart constructor.

Constructors

DescribeVpcClassicLinkResponse'

Fields

vpcs :: Maybe [VpcClassicLink]

The ClassicLink status of one or more VPCs.

httpStatus :: Int

The response's http status code.

newDescribeVpcClassicLinkResponse

Arguments

:: Int
$sel:httpStatus:DescribeVpcClassicLinkResponse'

-> DescribeVpcClassicLinkResponse

Create a value of DescribeVpcClassicLinkResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:vpcs:DescribeVpcClassicLinkResponse', describeVpcClassicLinkResponse_vpcs - The ClassicLink status of one or more VPCs.

$sel:httpStatus:DescribeVpcClassicLinkResponse', describeVpcClassicLinkResponse_httpStatus - The response's http status code.

Response Lenses

describeVpcClassicLinkResponse_vpcs :: Lens' DescribeVpcClassicLinkResponse (Maybe [VpcClassicLink])

The ClassicLink status of one or more VPCs.

describeVpcClassicLinkResponse_httpStatus :: Lens' DescribeVpcClassicLinkResponse Int

The response's http status code.