Amazonka.EC2.DescribeVpnGateways

Description

Describes one or more of your virtual private gateways.

For more information, see Amazon Web Services Site-to-Site VPN in the Amazon Web Services Site-to-Site VPN User Guide.

Creating a Request

data DescribeVpnGateways

Contains the parameters for DescribeVpnGateways.

See: newDescribeVpnGateways smart constructor.

Constructors

DescribeVpnGateways'

Fields

filters :: Maybe [Filter]

One or more filters.

amazon-side-asn - The Autonomous System Number (ASN) for the Amazon side of the gateway.

attachment.state - The current state of the attachment between the gateway and the VPC (attaching | attached | detaching | detached).

attachment.vpc-id - The ID of an attached VPC.

availability-zone - The Availability Zone for the virtual private gateway (if applicable).

state - The state of the virtual private gateway (pending | available | deleting | deleted).

tag:<key> - The key/value combination of a tag assigned to the resource. Use the tag key in the filter name and the tag value as the filter value. For example, to find all resources that have a tag with the key Owner and the value TeamA, specify tag:Owner for the filter name and TeamA for the filter value.

tag-key - The key of a tag assigned to the resource. Use this filter to find all resources assigned a tag with a specific key, regardless of the tag value.

type - The type of virtual private gateway. Currently the only supported type is ipsec.1.

vpn-gateway-id - The ID of the virtual private gateway.

vpnGatewayIds :: Maybe [Text]

One or more virtual private gateway IDs.

Default: Describes all your virtual private gateways.

dryRun :: Maybe Bool

Checks whether you have the required permissions for the action, without actually making the request, and provides an error response. If you have the required permissions, the error response is DryRunOperation. Otherwise, it is UnauthorizedOperation.

newDescribeVpnGateways :: DescribeVpnGateways

Create a value of DescribeVpnGateways with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:filters:DescribeVpnGateways', describeVpnGateways_filters - One or more filters.

amazon-side-asn - The Autonomous System Number (ASN) for the Amazon side of the gateway.

attachment.state - The current state of the attachment between the gateway and the VPC (attaching | attached | detaching | detached).

attachment.vpc-id - The ID of an attached VPC.

availability-zone - The Availability Zone for the virtual private gateway (if applicable).

state - The state of the virtual private gateway (pending | available | deleting | deleted).

tag:<key> - The key/value combination of a tag assigned to the resource. Use the tag key in the filter name and the tag value as the filter value. For example, to find all resources that have a tag with the key Owner and the value TeamA, specify tag:Owner for the filter name and TeamA for the filter value.

tag-key - The key of a tag assigned to the resource. Use this filter to find all resources assigned a tag with a specific key, regardless of the tag value.

type - The type of virtual private gateway. Currently the only supported type is ipsec.1.

vpn-gateway-id - The ID of the virtual private gateway.

$sel:vpnGatewayIds:DescribeVpnGateways', describeVpnGateways_vpnGatewayIds - One or more virtual private gateway IDs.

Default: Describes all your virtual private gateways.

$sel:dryRun:DescribeVpnGateways', describeVpnGateways_dryRun - Checks whether you have the required permissions for the action, without actually making the request, and provides an error response. If you have the required permissions, the error response is DryRunOperation. Otherwise, it is UnauthorizedOperation.

Request Lenses

describeVpnGateways_filters :: Lens' DescribeVpnGateways (Maybe [Filter])

One or more filters.

amazon-side-asn - The Autonomous System Number (ASN) for the Amazon side of the gateway.

attachment.state - The current state of the attachment between the gateway and the VPC (attaching | attached | detaching | detached).

attachment.vpc-id - The ID of an attached VPC.

availability-zone - The Availability Zone for the virtual private gateway (if applicable).

state - The state of the virtual private gateway (pending | available | deleting | deleted).

tag:<key> - The key/value combination of a tag assigned to the resource. Use the tag key in the filter name and the tag value as the filter value. For example, to find all resources that have a tag with the key Owner and the value TeamA, specify tag:Owner for the filter name and TeamA for the filter value.

tag-key - The key of a tag assigned to the resource. Use this filter to find all resources assigned a tag with a specific key, regardless of the tag value.

type - The type of virtual private gateway. Currently the only supported type is ipsec.1.

vpn-gateway-id - The ID of the virtual private gateway.

describeVpnGateways_vpnGatewayIds :: Lens' DescribeVpnGateways (Maybe [Text])

One or more virtual private gateway IDs.

Default: Describes all your virtual private gateways.

describeVpnGateways_dryRun :: Lens' DescribeVpnGateways (Maybe Bool)

Checks whether you have the required permissions for the action, without actually making the request, and provides an error response. If you have the required permissions, the error response is DryRunOperation. Otherwise, it is UnauthorizedOperation.

Destructuring the Response

data DescribeVpnGatewaysResponse

Contains the output of DescribeVpnGateways.

See: newDescribeVpnGatewaysResponse smart constructor.

Constructors

DescribeVpnGatewaysResponse'

Fields

vpnGateways :: Maybe [VpnGateway]

Information about one or more virtual private gateways.

httpStatus :: Int

The response's http status code.

newDescribeVpnGatewaysResponse

Arguments

:: Int
$sel:httpStatus:DescribeVpnGatewaysResponse'

-> DescribeVpnGatewaysResponse

Create a value of DescribeVpnGatewaysResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:vpnGateways:DescribeVpnGatewaysResponse', describeVpnGatewaysResponse_vpnGateways - Information about one or more virtual private gateways.

$sel:httpStatus:DescribeVpnGatewaysResponse', describeVpnGatewaysResponse_httpStatus - The response's http status code.

Response Lenses

describeVpnGatewaysResponse_vpnGateways :: Lens' DescribeVpnGatewaysResponse (Maybe [VpnGateway])

Information about one or more virtual private gateways.

describeVpnGatewaysResponse_httpStatus :: Lens' DescribeVpnGatewaysResponse Int

The response's http status code.