Amazonka.EC2.DetachClassicLinkVpc

Description

Unlinks (detaches) a linked EC2-Classic instance from a VPC. After the instance has been unlinked, the VPC security groups are no longer associated with it. An instance is automatically unlinked from a VPC when it's stopped.

Creating a Request

data DetachClassicLinkVpc

See: newDetachClassicLinkVpc smart constructor.

Constructors

DetachClassicLinkVpc'

Fields

dryRun :: Maybe Bool

Checks whether you have the required permissions for the action, without actually making the request, and provides an error response. If you have the required permissions, the error response is DryRunOperation. Otherwise, it is UnauthorizedOperation.

instanceId :: Text

The ID of the instance to unlink from the VPC.

vpcId :: Text

The ID of the VPC to which the instance is linked.

newDetachClassicLinkVpc

Arguments

:: Text
$sel:instanceId:DetachClassicLinkVpc'

-> Text
$sel:vpcId:DetachClassicLinkVpc'

-> DetachClassicLinkVpc

Create a value of DetachClassicLinkVpc with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:dryRun:DetachClassicLinkVpc', detachClassicLinkVpc_dryRun - Checks whether you have the required permissions for the action, without actually making the request, and provides an error response. If you have the required permissions, the error response is DryRunOperation. Otherwise, it is UnauthorizedOperation.

$sel:instanceId:DetachClassicLinkVpc', detachClassicLinkVpc_instanceId - The ID of the instance to unlink from the VPC.

$sel:vpcId:DetachClassicLinkVpc', detachClassicLinkVpc_vpcId - The ID of the VPC to which the instance is linked.

Request Lenses

detachClassicLinkVpc_dryRun :: Lens' DetachClassicLinkVpc (Maybe Bool)

Checks whether you have the required permissions for the action, without actually making the request, and provides an error response. If you have the required permissions, the error response is DryRunOperation. Otherwise, it is UnauthorizedOperation.

detachClassicLinkVpc_instanceId :: Lens' DetachClassicLinkVpc Text

The ID of the instance to unlink from the VPC.

detachClassicLinkVpc_vpcId :: Lens' DetachClassicLinkVpc Text

The ID of the VPC to which the instance is linked.

Destructuring the Response

data DetachClassicLinkVpcResponse

See: newDetachClassicLinkVpcResponse smart constructor.

Constructors

DetachClassicLinkVpcResponse'

Fields

return' :: Maybe Bool

Returns true if the request succeeds; otherwise, it returns an error.

httpStatus :: Int

The response's http status code.

newDetachClassicLinkVpcResponse

Arguments

:: Int
$sel:httpStatus:DetachClassicLinkVpcResponse'

-> DetachClassicLinkVpcResponse

Create a value of DetachClassicLinkVpcResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:return':DetachClassicLinkVpcResponse', detachClassicLinkVpcResponse_return - Returns true if the request succeeds; otherwise, it returns an error.

$sel:httpStatus:DetachClassicLinkVpcResponse', detachClassicLinkVpcResponse_httpStatus - The response's http status code.

Response Lenses

detachClassicLinkVpcResponse_return :: Lens' DetachClassicLinkVpcResponse (Maybe Bool)

Returns true if the request succeeds; otherwise, it returns an error.

detachClassicLinkVpcResponse_httpStatus :: Lens' DetachClassicLinkVpcResponse Int

The response's http status code.