Amazonka.EC2.DetachVolume

Description

Detaches an EBS volume from an instance. Make sure to unmount any file systems on the device within your operating system before detaching the volume. Failure to do so can result in the volume becoming stuck in the busy state while detaching. If this happens, detachment can be delayed indefinitely until you unmount the volume, force detachment, reboot the instance, or all three. If an EBS volume is the root device of an instance, it can't be detached while the instance is running. To detach the root volume, stop the instance first.

When a volume with an Amazon Web Services Marketplace product code is detached from an instance, the product code is no longer associated with the instance.

For more information, see Detach an Amazon EBS volume in the Amazon Elastic Compute Cloud User Guide.

Creating a Request

data DetachVolume

See: newDetachVolume smart constructor.

Constructors

DetachVolume'

Fields

instanceId :: Maybe Text

The ID of the instance. If you are detaching a Multi-Attach enabled volume, you must specify an instance ID.

force :: Maybe Bool

Forces detachment if the previous detachment attempt did not occur cleanly (for example, logging into an instance, unmounting the volume, and detaching normally). This option can lead to data loss or a corrupted file system. Use this option only as a last resort to detach a volume from a failed instance. The instance won't have an opportunity to flush file system caches or file system metadata. If you use this option, you must perform file system check and repair procedures.

device :: Maybe Text

The device name.

dryRun :: Maybe Bool

Checks whether you have the required permissions for the action, without actually making the request, and provides an error response. If you have the required permissions, the error response is DryRunOperation. Otherwise, it is UnauthorizedOperation.

volumeId :: Text

The ID of the volume.

newDetachVolume

Arguments

:: Text
$sel:volumeId:DetachVolume'

-> DetachVolume

Create a value of DetachVolume with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:instanceId:DetachVolume', detachVolume_instanceId - The ID of the instance. If you are detaching a Multi-Attach enabled volume, you must specify an instance ID.

$sel:force:DetachVolume', detachVolume_force - Forces detachment if the previous detachment attempt did not occur cleanly (for example, logging into an instance, unmounting the volume, and detaching normally). This option can lead to data loss or a corrupted file system. Use this option only as a last resort to detach a volume from a failed instance. The instance won't have an opportunity to flush file system caches or file system metadata. If you use this option, you must perform file system check and repair procedures.

$sel:device:DetachVolume', detachVolume_device - The device name.

$sel:dryRun:DetachVolume', detachVolume_dryRun - Checks whether you have the required permissions for the action, without actually making the request, and provides an error response. If you have the required permissions, the error response is DryRunOperation. Otherwise, it is UnauthorizedOperation.

$sel:volumeId:DetachVolume', detachVolume_volumeId - The ID of the volume.

Request Lenses

detachVolume_instanceId :: Lens' DetachVolume (Maybe Text)

The ID of the instance. If you are detaching a Multi-Attach enabled volume, you must specify an instance ID.

detachVolume_force :: Lens' DetachVolume (Maybe Bool)

Forces detachment if the previous detachment attempt did not occur cleanly (for example, logging into an instance, unmounting the volume, and detaching normally). This option can lead to data loss or a corrupted file system. Use this option only as a last resort to detach a volume from a failed instance. The instance won't have an opportunity to flush file system caches or file system metadata. If you use this option, you must perform file system check and repair procedures.

detachVolume_device :: Lens' DetachVolume (Maybe Text)

The device name.

detachVolume_dryRun :: Lens' DetachVolume (Maybe Bool)

Checks whether you have the required permissions for the action, without actually making the request, and provides an error response. If you have the required permissions, the error response is DryRunOperation. Otherwise, it is UnauthorizedOperation.

detachVolume_volumeId :: Lens' DetachVolume Text

The ID of the volume.

Destructuring the Response

data VolumeAttachment

Describes volume attachment details.

See: newVolumeAttachment smart constructor.

Constructors

VolumeAttachment'

Fields

instanceId :: Maybe Text

The ID of the instance.

deleteOnTermination :: Maybe Bool

Indicates whether the EBS volume is deleted on instance termination.

state :: Maybe VolumeAttachmentState

The attachment state of the volume.

device :: Maybe Text

The device name.

volumeId :: Maybe Text

The ID of the volume.

attachTime :: Maybe ISO8601

The time stamp when the attachment initiated.

newVolumeAttachment :: VolumeAttachment

Create a value of VolumeAttachment with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:instanceId:VolumeAttachment', volumeAttachment_instanceId - The ID of the instance.

$sel:deleteOnTermination:VolumeAttachment', volumeAttachment_deleteOnTermination - Indicates whether the EBS volume is deleted on instance termination.

$sel:state:VolumeAttachment', volumeAttachment_state - The attachment state of the volume.

$sel:device:VolumeAttachment', volumeAttachment_device - The device name.

$sel:volumeId:VolumeAttachment', volumeAttachment_volumeId - The ID of the volume.

$sel:attachTime:VolumeAttachment', volumeAttachment_attachTime - The time stamp when the attachment initiated.

Response Lenses

volumeAttachment_instanceId :: Lens' VolumeAttachment (Maybe Text)

The ID of the instance.

volumeAttachment_deleteOnTermination :: Lens' VolumeAttachment (Maybe Bool)

Indicates whether the EBS volume is deleted on instance termination.

volumeAttachment_state :: Lens' VolumeAttachment (Maybe VolumeAttachmentState)

The attachment state of the volume.

volumeAttachment_device :: Lens' VolumeAttachment (Maybe Text)

The device name.

volumeAttachment_volumeId :: Lens' VolumeAttachment (Maybe Text)

The ID of the volume.

volumeAttachment_attachTime :: Lens' VolumeAttachment (Maybe UTCTime)

The time stamp when the attachment initiated.