Amazonka.EC2.DisableVpcClassicLinkDnsSupport

Description

Disables ClassicLink DNS support for a VPC. If disabled, DNS hostnames resolve to public IP addresses when addressed between a linked EC2-Classic instance and instances in the VPC to which it's linked. For more information, see ClassicLink in the Amazon Elastic Compute Cloud User Guide.

You must specify a VPC ID in the request.

Creating a Request

data DisableVpcClassicLinkDnsSupport

See: newDisableVpcClassicLinkDnsSupport smart constructor.

Constructors

DisableVpcClassicLinkDnsSupport'

Fields

vpcId :: Maybe Text

The ID of the VPC.

newDisableVpcClassicLinkDnsSupport :: DisableVpcClassicLinkDnsSupport

Create a value of DisableVpcClassicLinkDnsSupport with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:vpcId:DisableVpcClassicLinkDnsSupport', disableVpcClassicLinkDnsSupport_vpcId - The ID of the VPC.

Request Lenses

disableVpcClassicLinkDnsSupport_vpcId :: Lens' DisableVpcClassicLinkDnsSupport (Maybe Text)

The ID of the VPC.

Destructuring the Response

data DisableVpcClassicLinkDnsSupportResponse

See: newDisableVpcClassicLinkDnsSupportResponse smart constructor.

Constructors

DisableVpcClassicLinkDnsSupportResponse'

Fields

return' :: Maybe Bool

Returns true if the request succeeds; otherwise, it returns an error.

httpStatus :: Int

The response's http status code.

newDisableVpcClassicLinkDnsSupportResponse

Arguments

:: Int
$sel:httpStatus:DisableVpcClassicLinkDnsSupportResponse'

-> DisableVpcClassicLinkDnsSupportResponse

Create a value of DisableVpcClassicLinkDnsSupportResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:return':DisableVpcClassicLinkDnsSupportResponse', disableVpcClassicLinkDnsSupportResponse_return - Returns true if the request succeeds; otherwise, it returns an error.

$sel:httpStatus:DisableVpcClassicLinkDnsSupportResponse', disableVpcClassicLinkDnsSupportResponse_httpStatus - The response's http status code.

Response Lenses

disableVpcClassicLinkDnsSupportResponse_return :: Lens' DisableVpcClassicLinkDnsSupportResponse (Maybe Bool)

Returns true if the request succeeds; otherwise, it returns an error.

disableVpcClassicLinkDnsSupportResponse_httpStatus :: Lens' DisableVpcClassicLinkDnsSupportResponse Int

The response's http status code.