Amazonka.EC2.ExportTransitGatewayRoutes

Description

Exports routes from the specified transit gateway route table to the specified S3 bucket. By default, all routes are exported. Alternatively, you can filter by CIDR range.

The routes are saved to the specified bucket in a JSON file. For more information, see Export Route Tables to Amazon S3 in Transit Gateways.

Creating a Request

data ExportTransitGatewayRoutes

See: newExportTransitGatewayRoutes smart constructor.

Constructors

ExportTransitGatewayRoutes'

Fields

filters :: Maybe [Filter]

One or more filters. The possible values are:

attachment.transit-gateway-attachment-id - The id of the transit gateway attachment.

attachment.resource-id - The resource id of the transit gateway attachment.

route-search.exact-match - The exact match of the specified filter.

route-search.longest-prefix-match - The longest prefix that matches the route.

route-search.subnet-of-match - The routes with a subnet that match the specified CIDR filter.

route-search.supernet-of-match - The routes with a CIDR that encompass the CIDR filter. For example, if you have 10.0.1.0/29 and 10.0.1.0/31 routes in your route table and you specify supernet-of-match as 10.0.1.0/30, then the result returns 10.0.1.0/29.

state - The state of the route (active | blackhole).

transit-gateway-route-destination-cidr-block - The CIDR range.

type - The type of route (propagated | static).

dryRun :: Maybe Bool

Checks whether you have the required permissions for the action, without actually making the request, and provides an error response. If you have the required permissions, the error response is DryRunOperation. Otherwise, it is UnauthorizedOperation.

transitGatewayRouteTableId :: Text

The ID of the route table.

s3Bucket :: Text

The name of the S3 bucket.

newExportTransitGatewayRoutes

Arguments

:: Text
$sel:transitGatewayRouteTableId:ExportTransitGatewayRoutes'

-> Text
$sel:s3Bucket:ExportTransitGatewayRoutes'

-> ExportTransitGatewayRoutes

Create a value of ExportTransitGatewayRoutes with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:filters:ExportTransitGatewayRoutes', exportTransitGatewayRoutes_filters - One or more filters. The possible values are:

attachment.transit-gateway-attachment-id - The id of the transit gateway attachment.

attachment.resource-id - The resource id of the transit gateway attachment.

route-search.exact-match - The exact match of the specified filter.

route-search.longest-prefix-match - The longest prefix that matches the route.

route-search.subnet-of-match - The routes with a subnet that match the specified CIDR filter.

route-search.supernet-of-match - The routes with a CIDR that encompass the CIDR filter. For example, if you have 10.0.1.0/29 and 10.0.1.0/31 routes in your route table and you specify supernet-of-match as 10.0.1.0/30, then the result returns 10.0.1.0/29.

state - The state of the route (active | blackhole).

transit-gateway-route-destination-cidr-block - The CIDR range.

type - The type of route (propagated | static).

$sel:dryRun:ExportTransitGatewayRoutes', exportTransitGatewayRoutes_dryRun - Checks whether you have the required permissions for the action, without actually making the request, and provides an error response. If you have the required permissions, the error response is DryRunOperation. Otherwise, it is UnauthorizedOperation.

$sel:transitGatewayRouteTableId:ExportTransitGatewayRoutes', exportTransitGatewayRoutes_transitGatewayRouteTableId - The ID of the route table.

$sel:s3Bucket:ExportTransitGatewayRoutes', exportTransitGatewayRoutes_s3Bucket - The name of the S3 bucket.

Request Lenses

exportTransitGatewayRoutes_filters :: Lens' ExportTransitGatewayRoutes (Maybe [Filter])

One or more filters. The possible values are:

attachment.transit-gateway-attachment-id - The id of the transit gateway attachment.

attachment.resource-id - The resource id of the transit gateway attachment.

route-search.exact-match - The exact match of the specified filter.

route-search.longest-prefix-match - The longest prefix that matches the route.

route-search.subnet-of-match - The routes with a subnet that match the specified CIDR filter.

route-search.supernet-of-match - The routes with a CIDR that encompass the CIDR filter. For example, if you have 10.0.1.0/29 and 10.0.1.0/31 routes in your route table and you specify supernet-of-match as 10.0.1.0/30, then the result returns 10.0.1.0/29.

state - The state of the route (active | blackhole).

transit-gateway-route-destination-cidr-block - The CIDR range.

type - The type of route (propagated | static).

exportTransitGatewayRoutes_dryRun :: Lens' ExportTransitGatewayRoutes (Maybe Bool)

Checks whether you have the required permissions for the action, without actually making the request, and provides an error response. If you have the required permissions, the error response is DryRunOperation. Otherwise, it is UnauthorizedOperation.

exportTransitGatewayRoutes_transitGatewayRouteTableId :: Lens' ExportTransitGatewayRoutes Text

The ID of the route table.

exportTransitGatewayRoutes_s3Bucket :: Lens' ExportTransitGatewayRoutes Text

The name of the S3 bucket.

Destructuring the Response

data ExportTransitGatewayRoutesResponse

See: newExportTransitGatewayRoutesResponse smart constructor.

Constructors

ExportTransitGatewayRoutesResponse'

Fields

s3Location :: Maybe Text

The URL of the exported file in Amazon S3. For example, s3://bucket_name/VPCTransitGateway/TransitGatewayRouteTables/file_name.

httpStatus :: Int

The response's http status code.

newExportTransitGatewayRoutesResponse

Arguments

:: Int
$sel:httpStatus:ExportTransitGatewayRoutesResponse'

-> ExportTransitGatewayRoutesResponse

Create a value of ExportTransitGatewayRoutesResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:s3Location:ExportTransitGatewayRoutesResponse', exportTransitGatewayRoutesResponse_s3Location - The URL of the exported file in Amazon S3. For example, s3://bucket_name/VPCTransitGateway/TransitGatewayRouteTables/file_name.

$sel:httpStatus:ExportTransitGatewayRoutesResponse', exportTransitGatewayRoutesResponse_httpStatus - The response's http status code.

Response Lenses

exportTransitGatewayRoutesResponse_s3Location :: Lens' ExportTransitGatewayRoutesResponse (Maybe Text)

The URL of the exported file in Amazon S3. For example, s3://bucket_name/VPCTransitGateway/TransitGatewayRouteTables/file_name.

exportTransitGatewayRoutesResponse_httpStatus :: Lens' ExportTransitGatewayRoutesResponse Int

The response's http status code.