Amazonka.EC2.GetTransitGatewayRouteTablePropagations

Description

Gets information about the route table propagations for the specified transit gateway route table.

This operation returns paginated results.

Creating a Request

data GetTransitGatewayRouteTablePropagations

See: newGetTransitGatewayRouteTablePropagations smart constructor.

Constructors

GetTransitGatewayRouteTablePropagations'

Fields

filters :: Maybe [Filter]

One or more filters. The possible values are:

resource-id - The ID of the resource.

resource-type - The resource type. Valid values are vpc | vpn | direct-connect-gateway | peering | connect.

transit-gateway-attachment-id - The ID of the attachment.

nextToken :: Maybe Text

The token for the next page of results.

dryRun :: Maybe Bool

Checks whether you have the required permissions for the action, without actually making the request, and provides an error response. If you have the required permissions, the error response is DryRunOperation. Otherwise, it is UnauthorizedOperation.

maxResults :: Maybe Natural

The maximum number of results to return with a single call. To retrieve the remaining results, make another call with the returned nextToken value.

transitGatewayRouteTableId :: Text

The ID of the transit gateway route table.

newGetTransitGatewayRouteTablePropagations

Arguments

:: Text
$sel:transitGatewayRouteTableId:GetTransitGatewayRouteTablePropagations'

-> GetTransitGatewayRouteTablePropagations

Create a value of GetTransitGatewayRouteTablePropagations with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:filters:GetTransitGatewayRouteTablePropagations', getTransitGatewayRouteTablePropagations_filters - One or more filters. The possible values are:

resource-id - The ID of the resource.

resource-type - The resource type. Valid values are vpc | vpn | direct-connect-gateway | peering | connect.

transit-gateway-attachment-id - The ID of the attachment.

$sel:nextToken:GetTransitGatewayRouteTablePropagations', getTransitGatewayRouteTablePropagations_nextToken - The token for the next page of results.

$sel:dryRun:GetTransitGatewayRouteTablePropagations', getTransitGatewayRouteTablePropagations_dryRun - Checks whether you have the required permissions for the action, without actually making the request, and provides an error response. If you have the required permissions, the error response is DryRunOperation. Otherwise, it is UnauthorizedOperation.

$sel:maxResults:GetTransitGatewayRouteTablePropagations', getTransitGatewayRouteTablePropagations_maxResults - The maximum number of results to return with a single call. To retrieve the remaining results, make another call with the returned nextToken value.

$sel:transitGatewayRouteTableId:GetTransitGatewayRouteTablePropagations', getTransitGatewayRouteTablePropagations_transitGatewayRouteTableId - The ID of the transit gateway route table.

Request Lenses

getTransitGatewayRouteTablePropagations_filters :: Lens' GetTransitGatewayRouteTablePropagations (Maybe [Filter])

One or more filters. The possible values are:

resource-id - The ID of the resource.

resource-type - The resource type. Valid values are vpc | vpn | direct-connect-gateway | peering | connect.

transit-gateway-attachment-id - The ID of the attachment.

getTransitGatewayRouteTablePropagations_nextToken :: Lens' GetTransitGatewayRouteTablePropagations (Maybe Text)

The token for the next page of results.

getTransitGatewayRouteTablePropagations_dryRun :: Lens' GetTransitGatewayRouteTablePropagations (Maybe Bool)

Checks whether you have the required permissions for the action, without actually making the request, and provides an error response. If you have the required permissions, the error response is DryRunOperation. Otherwise, it is UnauthorizedOperation.

getTransitGatewayRouteTablePropagations_maxResults :: Lens' GetTransitGatewayRouteTablePropagations (Maybe Natural)

The maximum number of results to return with a single call. To retrieve the remaining results, make another call with the returned nextToken value.

getTransitGatewayRouteTablePropagations_transitGatewayRouteTableId :: Lens' GetTransitGatewayRouteTablePropagations Text

The ID of the transit gateway route table.

Destructuring the Response

data GetTransitGatewayRouteTablePropagationsResponse

See: newGetTransitGatewayRouteTablePropagationsResponse smart constructor.

Constructors

GetTransitGatewayRouteTablePropagationsResponse'

Fields

transitGatewayRouteTablePropagations :: Maybe [TransitGatewayRouteTablePropagation]

Information about the route table propagations.

nextToken :: Maybe Text

The token to use to retrieve the next page of results. This value is null when there are no more results to return.

httpStatus :: Int

The response's http status code.

newGetTransitGatewayRouteTablePropagationsResponse

Arguments

:: Int
$sel:httpStatus:GetTransitGatewayRouteTablePropagationsResponse'

-> GetTransitGatewayRouteTablePropagationsResponse

Create a value of GetTransitGatewayRouteTablePropagationsResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:transitGatewayRouteTablePropagations:GetTransitGatewayRouteTablePropagationsResponse', getTransitGatewayRouteTablePropagationsResponse_transitGatewayRouteTablePropagations - Information about the route table propagations.

$sel:nextToken:GetTransitGatewayRouteTablePropagations', getTransitGatewayRouteTablePropagationsResponse_nextToken - The token to use to retrieve the next page of results. This value is null when there are no more results to return.

$sel:httpStatus:GetTransitGatewayRouteTablePropagationsResponse', getTransitGatewayRouteTablePropagationsResponse_httpStatus - The response's http status code.

Response Lenses

getTransitGatewayRouteTablePropagationsResponse_transitGatewayRouteTablePropagations :: Lens' GetTransitGatewayRouteTablePropagationsResponse (Maybe [TransitGatewayRouteTablePropagation])

Information about the route table propagations.

getTransitGatewayRouteTablePropagationsResponse_nextToken :: Lens' GetTransitGatewayRouteTablePropagationsResponse (Maybe Text)

The token to use to retrieve the next page of results. This value is null when there are no more results to return.

getTransitGatewayRouteTablePropagationsResponse_httpStatus :: Lens' GetTransitGatewayRouteTablePropagationsResponse Int

The response's http status code.