Amazonka.EC2.GetVpnConnectionDeviceTypes

Description

Obtain a list of customer gateway devices for which sample configuration files can be provided. The request has no additional parameters. You can also see the list of device types with sample configuration files available under Your customer gateway device in the Amazon Web Services Site-to-Site VPN User Guide.

This operation returns paginated results.

Creating a Request

data GetVpnConnectionDeviceTypes

See: newGetVpnConnectionDeviceTypes smart constructor.

Constructors

GetVpnConnectionDeviceTypes'

Fields

nextToken :: Maybe Text

The NextToken value returned from a previous paginated GetVpnConnectionDeviceTypes request where MaxResults was used and the results exceeded the value of that parameter. Pagination continues from the end of the previous results that returned the NextToken value. This value is null when there are no more results to return.

dryRun :: Maybe Bool

Checks whether you have the required permissions for the action, without actually making the request, and provides an error response. If you have the required permissions, the error response is DryRunOperation. Otherwise, it is UnauthorizedOperation.

maxResults :: Maybe Natural

The maximum number of results returned by GetVpnConnectionDeviceTypes in paginated output. When this parameter is used, GetVpnConnectionDeviceTypes only returns MaxResults results in a single page along with a NextToken response element. The remaining results of the initial request can be seen by sending another GetVpnConnectionDeviceTypes request with the returned NextToken value. This value can be between 200 and 1000. If this parameter is not used, then GetVpnConnectionDeviceTypes returns all results.

newGetVpnConnectionDeviceTypes :: GetVpnConnectionDeviceTypes

Create a value of GetVpnConnectionDeviceTypes with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:nextToken:GetVpnConnectionDeviceTypes', getVpnConnectionDeviceTypes_nextToken - The NextToken value returned from a previous paginated GetVpnConnectionDeviceTypes request where MaxResults was used and the results exceeded the value of that parameter. Pagination continues from the end of the previous results that returned the NextToken value. This value is null when there are no more results to return.

$sel:dryRun:GetVpnConnectionDeviceTypes', getVpnConnectionDeviceTypes_dryRun - Checks whether you have the required permissions for the action, without actually making the request, and provides an error response. If you have the required permissions, the error response is DryRunOperation. Otherwise, it is UnauthorizedOperation.

$sel:maxResults:GetVpnConnectionDeviceTypes', getVpnConnectionDeviceTypes_maxResults - The maximum number of results returned by GetVpnConnectionDeviceTypes in paginated output. When this parameter is used, GetVpnConnectionDeviceTypes only returns MaxResults results in a single page along with a NextToken response element. The remaining results of the initial request can be seen by sending another GetVpnConnectionDeviceTypes request with the returned NextToken value. This value can be between 200 and 1000. If this parameter is not used, then GetVpnConnectionDeviceTypes returns all results.

Request Lenses

getVpnConnectionDeviceTypes_nextToken :: Lens' GetVpnConnectionDeviceTypes (Maybe Text)

The NextToken value returned from a previous paginated GetVpnConnectionDeviceTypes request where MaxResults was used and the results exceeded the value of that parameter. Pagination continues from the end of the previous results that returned the NextToken value. This value is null when there are no more results to return.

getVpnConnectionDeviceTypes_dryRun :: Lens' GetVpnConnectionDeviceTypes (Maybe Bool)

Checks whether you have the required permissions for the action, without actually making the request, and provides an error response. If you have the required permissions, the error response is DryRunOperation. Otherwise, it is UnauthorizedOperation.

getVpnConnectionDeviceTypes_maxResults :: Lens' GetVpnConnectionDeviceTypes (Maybe Natural)

The maximum number of results returned by GetVpnConnectionDeviceTypes in paginated output. When this parameter is used, GetVpnConnectionDeviceTypes only returns MaxResults results in a single page along with a NextToken response element. The remaining results of the initial request can be seen by sending another GetVpnConnectionDeviceTypes request with the returned NextToken value. This value can be between 200 and 1000. If this parameter is not used, then GetVpnConnectionDeviceTypes returns all results.

Destructuring the Response

data GetVpnConnectionDeviceTypesResponse

See: newGetVpnConnectionDeviceTypesResponse smart constructor.

Constructors

GetVpnConnectionDeviceTypesResponse'

Fields

vpnConnectionDeviceTypes :: Maybe [VpnConnectionDeviceType]

List of customer gateway devices that have a sample configuration file available for use.

nextToken :: Maybe Text

The NextToken value to include in a future GetVpnConnectionDeviceTypes request. When the results of a GetVpnConnectionDeviceTypes request exceed MaxResults, this value can be used to retrieve the next page of results. This value is null when there are no more results to return.

httpStatus :: Int

The response's http status code.

newGetVpnConnectionDeviceTypesResponse

Arguments

:: Int
$sel:httpStatus:GetVpnConnectionDeviceTypesResponse'

-> GetVpnConnectionDeviceTypesResponse

Create a value of GetVpnConnectionDeviceTypesResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:vpnConnectionDeviceTypes:GetVpnConnectionDeviceTypesResponse', getVpnConnectionDeviceTypesResponse_vpnConnectionDeviceTypes - List of customer gateway devices that have a sample configuration file available for use.

$sel:nextToken:GetVpnConnectionDeviceTypes', getVpnConnectionDeviceTypesResponse_nextToken - The NextToken value to include in a future GetVpnConnectionDeviceTypes request. When the results of a GetVpnConnectionDeviceTypes request exceed MaxResults, this value can be used to retrieve the next page of results. This value is null when there are no more results to return.

$sel:httpStatus:GetVpnConnectionDeviceTypesResponse', getVpnConnectionDeviceTypesResponse_httpStatus - The response's http status code.

Response Lenses

getVpnConnectionDeviceTypesResponse_vpnConnectionDeviceTypes :: Lens' GetVpnConnectionDeviceTypesResponse (Maybe [VpnConnectionDeviceType])

List of customer gateway devices that have a sample configuration file available for use.

getVpnConnectionDeviceTypesResponse_nextToken :: Lens' GetVpnConnectionDeviceTypesResponse (Maybe Text)

The NextToken value to include in a future GetVpnConnectionDeviceTypes request. When the results of a GetVpnConnectionDeviceTypes request exceed MaxResults, this value can be used to retrieve the next page of results. This value is null when there are no more results to return.

getVpnConnectionDeviceTypesResponse_httpStatus :: Lens' GetVpnConnectionDeviceTypesResponse Int

The response's http status code.