Amazonka.EC2.ImportClientVpnClientCertificateRevocationList

Description

Uploads a client certificate revocation list to the specified Client VPN endpoint. Uploading a client certificate revocation list overwrites the existing client certificate revocation list.

Uploading a client certificate revocation list resets existing client connections.

Creating a Request

data ImportClientVpnClientCertificateRevocationList

See: newImportClientVpnClientCertificateRevocationList smart constructor.

Constructors

ImportClientVpnClientCertificateRevocationList'

Fields

dryRun :: Maybe Bool

Checks whether you have the required permissions for the action, without actually making the request, and provides an error response. If you have the required permissions, the error response is DryRunOperation. Otherwise, it is UnauthorizedOperation.

clientVpnEndpointId :: Text

The ID of the Client VPN endpoint to which the client certificate revocation list applies.

certificateRevocationList :: Text

The client certificate revocation list file. For more information, see Generate a Client Certificate Revocation List in the Client VPN Administrator Guide.

newImportClientVpnClientCertificateRevocationList

Arguments

:: Text
$sel:clientVpnEndpointId:ImportClientVpnClientCertificateRevocationList'

-> Text
$sel:certificateRevocationList:ImportClientVpnClientCertificateRevocationList'

-> ImportClientVpnClientCertificateRevocationList

Create a value of ImportClientVpnClientCertificateRevocationList with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:dryRun:ImportClientVpnClientCertificateRevocationList', importClientVpnClientCertificateRevocationList_dryRun - Checks whether you have the required permissions for the action, without actually making the request, and provides an error response. If you have the required permissions, the error response is DryRunOperation. Otherwise, it is UnauthorizedOperation.

$sel:clientVpnEndpointId:ImportClientVpnClientCertificateRevocationList', importClientVpnClientCertificateRevocationList_clientVpnEndpointId - The ID of the Client VPN endpoint to which the client certificate revocation list applies.

$sel:certificateRevocationList:ImportClientVpnClientCertificateRevocationList', importClientVpnClientCertificateRevocationList_certificateRevocationList - The client certificate revocation list file. For more information, see Generate a Client Certificate Revocation List in the Client VPN Administrator Guide.

Request Lenses

importClientVpnClientCertificateRevocationList_dryRun :: Lens' ImportClientVpnClientCertificateRevocationList (Maybe Bool)

Checks whether you have the required permissions for the action, without actually making the request, and provides an error response. If you have the required permissions, the error response is DryRunOperation. Otherwise, it is UnauthorizedOperation.

importClientVpnClientCertificateRevocationList_clientVpnEndpointId :: Lens' ImportClientVpnClientCertificateRevocationList Text

The ID of the Client VPN endpoint to which the client certificate revocation list applies.

importClientVpnClientCertificateRevocationList_certificateRevocationList :: Lens' ImportClientVpnClientCertificateRevocationList Text

The client certificate revocation list file. For more information, see Generate a Client Certificate Revocation List in the Client VPN Administrator Guide.

Destructuring the Response

data ImportClientVpnClientCertificateRevocationListResponse

See: newImportClientVpnClientCertificateRevocationListResponse smart constructor.

Constructors

ImportClientVpnClientCertificateRevocationListResponse'

Fields

return' :: Maybe Bool

Returns true if the request succeeds; otherwise, it returns an error.

httpStatus :: Int

The response's http status code.

newImportClientVpnClientCertificateRevocationListResponse

Arguments

:: Int
$sel:httpStatus:ImportClientVpnClientCertificateRevocationListResponse'

-> ImportClientVpnClientCertificateRevocationListResponse

Create a value of ImportClientVpnClientCertificateRevocationListResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:return':ImportClientVpnClientCertificateRevocationListResponse', importClientVpnClientCertificateRevocationListResponse_return - Returns true if the request succeeds; otherwise, it returns an error.

$sel:httpStatus:ImportClientVpnClientCertificateRevocationListResponse', importClientVpnClientCertificateRevocationListResponse_httpStatus - The response's http status code.

Response Lenses

importClientVpnClientCertificateRevocationListResponse_return :: Lens' ImportClientVpnClientCertificateRevocationListResponse (Maybe Bool)

Returns true if the request succeeds; otherwise, it returns an error.

importClientVpnClientCertificateRevocationListResponse_httpStatus :: Lens' ImportClientVpnClientCertificateRevocationListResponse Int

The response's http status code.