Amazonka.EC2.ImportImage

Description

Import single or multi-volume disk images or EBS snapshots into an Amazon Machine Image (AMI).

For more information, see Importing a VM as an image using VM Import/Export in the VM Import/Export User Guide.

Creating a Request

data ImportImage

See: newImportImage smart constructor.

Constructors

ImportImage'

Fields

hypervisor :: Maybe Text

The target hypervisor platform.

Valid values: xen

platform :: Maybe Text

The operating system of the virtual machine.

Valid values: Windows | Linux

clientToken :: Maybe Text

The token to enable idempotency for VM import requests.

licenseSpecifications :: Maybe [ImportImageLicenseConfigurationRequest]

The ARNs of the license configurations.

usageOperation :: Maybe Text

The usage operation value. For more information, see Licensing options in the VM Import/Export User Guide.

licenseType :: Maybe Text

The license type to be used for the Amazon Machine Image (AMI) after importing.

By default, we detect the source-system operating system (OS) and apply the appropriate license. Specify AWS to replace the source-system license with an Amazon Web Services license, if appropriate. Specify BYOL to retain the source-system license, if appropriate.

To use BYOL, you must have existing licenses with rights to use these licenses in a third party cloud, such as Amazon Web Services. For more information, see Prerequisites in the VM Import/Export User Guide.

roleName :: Maybe Text

The name of the role to use when not using the default role, 'vmimport'.

encrypted :: Maybe Bool

Specifies whether the destination AMI of the imported image should be encrypted. The default KMS key for EBS is used unless you specify a non-default KMS key using KmsKeyId. For more information, see Amazon EBS Encryption in the Amazon Elastic Compute Cloud User Guide.

tagSpecifications :: Maybe [TagSpecification]

The tags to apply to the import image task during creation.

kmsKeyId :: Maybe Text

An identifier for the symmetric KMS key to use when creating the encrypted AMI. This parameter is only required if you want to use a non-default KMS key; if this parameter is not specified, the default KMS key for EBS is used. If a KmsKeyId is specified, the Encrypted flag must also be set.

The KMS key identifier may be provided in any of the following formats:

Key ID

Key alias. The alias ARN contains the arn:aws:kms namespace, followed by the Region of the key, the Amazon Web Services account ID of the key owner, the alias namespace, and then the key alias. For example, arn:aws:kms:us-east-1:012345678910:alias/ExampleAlias.

ARN using key ID. The ID ARN contains the arn:aws:kms namespace, followed by the Region of the key, the Amazon Web Services account ID of the key owner, the key namespace, and then the key ID. For example, arn:aws:kms:us-east-1:012345678910:key/abcd1234-a123-456a-a12b-a123b4cd56ef.

ARN using key alias. The alias ARN contains the arn:aws:kms namespace, followed by the Region of the key, the Amazon Web Services account ID of the key owner, the alias namespace, and then the key alias. For example, arn:aws:kms:us-east-1:012345678910:alias/ExampleAlias.

Amazon Web Services parses KmsKeyId asynchronously, meaning that the action you call may appear to complete even though you provided an invalid identifier. This action will eventually report failure.

The specified KMS key must exist in the Region that the AMI is being copied to.

Amazon EBS does not support asymmetric KMS keys.

bootMode :: Maybe BootModeValues

The boot mode of the virtual machine.

architecture :: Maybe Text

The architecture of the virtual machine.

Valid values: i386 | x86_64 | arm64

description :: Maybe Text

A description string for the import image task.

dryRun :: Maybe Bool

Checks whether you have the required permissions for the action, without actually making the request, and provides an error response. If you have the required permissions, the error response is DryRunOperation. Otherwise, it is UnauthorizedOperation.

clientData :: Maybe ClientData

The client-specific data.

diskContainers :: Maybe [ImageDiskContainer]

Information about the disk containers.

newImportImage :: ImportImage

Create a value of ImportImage with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:hypervisor:ImportImage', importImage_hypervisor - The target hypervisor platform.

Valid values: xen

$sel:platform:ImportImage', importImage_platform - The operating system of the virtual machine.

Valid values: Windows | Linux

$sel:clientToken:ImportImage', importImage_clientToken - The token to enable idempotency for VM import requests.

$sel:licenseSpecifications:ImportImage', importImage_licenseSpecifications - The ARNs of the license configurations.

$sel:usageOperation:ImportImage', importImage_usageOperation - The usage operation value. For more information, see Licensing options in the VM Import/Export User Guide.

$sel:licenseType:ImportImage', importImage_licenseType - The license type to be used for the Amazon Machine Image (AMI) after importing.

By default, we detect the source-system operating system (OS) and apply the appropriate license. Specify AWS to replace the source-system license with an Amazon Web Services license, if appropriate. Specify BYOL to retain the source-system license, if appropriate.

To use BYOL, you must have existing licenses with rights to use these licenses in a third party cloud, such as Amazon Web Services. For more information, see Prerequisites in the VM Import/Export User Guide.

$sel:roleName:ImportImage', importImage_roleName - The name of the role to use when not using the default role, 'vmimport'.

$sel:encrypted:ImportImage', importImage_encrypted - Specifies whether the destination AMI of the imported image should be encrypted. The default KMS key for EBS is used unless you specify a non-default KMS key using KmsKeyId. For more information, see Amazon EBS Encryption in the Amazon Elastic Compute Cloud User Guide.

$sel:tagSpecifications:ImportImage', importImage_tagSpecifications - The tags to apply to the import image task during creation.

$sel:kmsKeyId:ImportImage', importImage_kmsKeyId - An identifier for the symmetric KMS key to use when creating the encrypted AMI. This parameter is only required if you want to use a non-default KMS key; if this parameter is not specified, the default KMS key for EBS is used. If a KmsKeyId is specified, the Encrypted flag must also be set.

The KMS key identifier may be provided in any of the following formats:

Key ID

Key alias. The alias ARN contains the arn:aws:kms namespace, followed by the Region of the key, the Amazon Web Services account ID of the key owner, the alias namespace, and then the key alias. For example, arn:aws:kms:us-east-1:012345678910:alias/ExampleAlias.

ARN using key ID. The ID ARN contains the arn:aws:kms namespace, followed by the Region of the key, the Amazon Web Services account ID of the key owner, the key namespace, and then the key ID. For example, arn:aws:kms:us-east-1:012345678910:key/abcd1234-a123-456a-a12b-a123b4cd56ef.

ARN using key alias. The alias ARN contains the arn:aws:kms namespace, followed by the Region of the key, the Amazon Web Services account ID of the key owner, the alias namespace, and then the key alias. For example, arn:aws:kms:us-east-1:012345678910:alias/ExampleAlias.

Amazon Web Services parses KmsKeyId asynchronously, meaning that the action you call may appear to complete even though you provided an invalid identifier. This action will eventually report failure.

The specified KMS key must exist in the Region that the AMI is being copied to.

Amazon EBS does not support asymmetric KMS keys.

$sel:bootMode:ImportImage', importImage_bootMode - The boot mode of the virtual machine.

$sel:architecture:ImportImage', importImage_architecture - The architecture of the virtual machine.

Valid values: i386 | x86_64 | arm64

$sel:description:ImportImage', importImage_description - A description string for the import image task.

$sel:dryRun:ImportImage', importImage_dryRun - Checks whether you have the required permissions for the action, without actually making the request, and provides an error response. If you have the required permissions, the error response is DryRunOperation. Otherwise, it is UnauthorizedOperation.

$sel:clientData:ImportImage', importImage_clientData - The client-specific data.

$sel:diskContainers:ImportImage', importImage_diskContainers - Information about the disk containers.

Request Lenses

importImage_hypervisor :: Lens' ImportImage (Maybe Text)

The target hypervisor platform.

Valid values: xen

importImage_platform :: Lens' ImportImage (Maybe Text)

The operating system of the virtual machine.

Valid values: Windows | Linux

importImage_clientToken :: Lens' ImportImage (Maybe Text)

The token to enable idempotency for VM import requests.

importImage_licenseSpecifications :: Lens' ImportImage (Maybe [ImportImageLicenseConfigurationRequest])

The ARNs of the license configurations.

importImage_usageOperation :: Lens' ImportImage (Maybe Text)

The usage operation value. For more information, see Licensing options in the VM Import/Export User Guide.

importImage_licenseType :: Lens' ImportImage (Maybe Text)

The license type to be used for the Amazon Machine Image (AMI) after importing.

By default, we detect the source-system operating system (OS) and apply the appropriate license. Specify AWS to replace the source-system license with an Amazon Web Services license, if appropriate. Specify BYOL to retain the source-system license, if appropriate.

To use BYOL, you must have existing licenses with rights to use these licenses in a third party cloud, such as Amazon Web Services. For more information, see Prerequisites in the VM Import/Export User Guide.

importImage_roleName :: Lens' ImportImage (Maybe Text)

The name of the role to use when not using the default role, 'vmimport'.

importImage_encrypted :: Lens' ImportImage (Maybe Bool)

Specifies whether the destination AMI of the imported image should be encrypted. The default KMS key for EBS is used unless you specify a non-default KMS key using KmsKeyId. For more information, see Amazon EBS Encryption in the Amazon Elastic Compute Cloud User Guide.

importImage_tagSpecifications :: Lens' ImportImage (Maybe [TagSpecification])

The tags to apply to the import image task during creation.

importImage_kmsKeyId :: Lens' ImportImage (Maybe Text)

An identifier for the symmetric KMS key to use when creating the encrypted AMI. This parameter is only required if you want to use a non-default KMS key; if this parameter is not specified, the default KMS key for EBS is used. If a KmsKeyId is specified, the Encrypted flag must also be set.

The KMS key identifier may be provided in any of the following formats:

Key ID

Key alias. The alias ARN contains the arn:aws:kms namespace, followed by the Region of the key, the Amazon Web Services account ID of the key owner, the alias namespace, and then the key alias. For example, arn:aws:kms:us-east-1:012345678910:alias/ExampleAlias.

ARN using key ID. The ID ARN contains the arn:aws:kms namespace, followed by the Region of the key, the Amazon Web Services account ID of the key owner, the key namespace, and then the key ID. For example, arn:aws:kms:us-east-1:012345678910:key/abcd1234-a123-456a-a12b-a123b4cd56ef.

ARN using key alias. The alias ARN contains the arn:aws:kms namespace, followed by the Region of the key, the Amazon Web Services account ID of the key owner, the alias namespace, and then the key alias. For example, arn:aws:kms:us-east-1:012345678910:alias/ExampleAlias.

Amazon Web Services parses KmsKeyId asynchronously, meaning that the action you call may appear to complete even though you provided an invalid identifier. This action will eventually report failure.

The specified KMS key must exist in the Region that the AMI is being copied to.

Amazon EBS does not support asymmetric KMS keys.

importImage_bootMode :: Lens' ImportImage (Maybe BootModeValues)

The boot mode of the virtual machine.

importImage_architecture :: Lens' ImportImage (Maybe Text)

The architecture of the virtual machine.

Valid values: i386 | x86_64 | arm64

importImage_description :: Lens' ImportImage (Maybe Text)

A description string for the import image task.

importImage_dryRun :: Lens' ImportImage (Maybe Bool)

Checks whether you have the required permissions for the action, without actually making the request, and provides an error response. If you have the required permissions, the error response is DryRunOperation. Otherwise, it is UnauthorizedOperation.

importImage_clientData :: Lens' ImportImage (Maybe ClientData)

The client-specific data.

importImage_diskContainers :: Lens' ImportImage (Maybe [ImageDiskContainer])

Information about the disk containers.

Destructuring the Response

data ImportImageResponse

See: newImportImageResponse smart constructor.

Constructors

ImportImageResponse'

Fields

status :: Maybe Text

A brief status of the task.

hypervisor :: Maybe Text

The target hypervisor of the import task.

platform :: Maybe Text

The operating system of the virtual machine.

progress :: Maybe Text

The progress of the task.

licenseSpecifications :: Maybe [ImportImageLicenseConfigurationResponse]

The ARNs of the license configurations.

usageOperation :: Maybe Text

The usage operation value.

licenseType :: Maybe Text

The license type of the virtual machine.

snapshotDetails :: Maybe [SnapshotDetail]

Information about the snapshots.

encrypted :: Maybe Bool

Indicates whether the AMI is encrypted.

kmsKeyId :: Maybe Text

The identifier for the symmetric KMS key that was used to create the encrypted AMI.

statusMessage :: Maybe Text

A detailed status message of the import task.

imageId :: Maybe Text

The ID of the Amazon Machine Image (AMI) created by the import task.

importTaskId :: Maybe Text

The task ID of the import image task.

architecture :: Maybe Text

The architecture of the virtual machine.

description :: Maybe Text

A description of the import task.

tags :: Maybe [Tag]

Any tags assigned to the import image task.

httpStatus :: Int

The response's http status code.

newImportImageResponse

Arguments

:: Int
$sel:httpStatus:ImportImageResponse'

-> ImportImageResponse

Create a value of ImportImageResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:status:ImportImageResponse', importImageResponse_status - A brief status of the task.

$sel:hypervisor:ImportImage', importImageResponse_hypervisor - The target hypervisor of the import task.

$sel:platform:ImportImage', importImageResponse_platform - The operating system of the virtual machine.

$sel:progress:ImportImageResponse', importImageResponse_progress - The progress of the task.

$sel:licenseSpecifications:ImportImage', importImageResponse_licenseSpecifications - The ARNs of the license configurations.

$sel:usageOperation:ImportImage', importImageResponse_usageOperation - The usage operation value.

$sel:licenseType:ImportImage', importImageResponse_licenseType - The license type of the virtual machine.

$sel:snapshotDetails:ImportImageResponse', importImageResponse_snapshotDetails - Information about the snapshots.

$sel:encrypted:ImportImage', importImageResponse_encrypted - Indicates whether the AMI is encrypted.

$sel:kmsKeyId:ImportImage', importImageResponse_kmsKeyId - The identifier for the symmetric KMS key that was used to create the encrypted AMI.

$sel:statusMessage:ImportImageResponse', importImageResponse_statusMessage - A detailed status message of the import task.

$sel:imageId:ImportImageResponse', importImageResponse_imageId - The ID of the Amazon Machine Image (AMI) created by the import task.

$sel:importTaskId:ImportImageResponse', importImageResponse_importTaskId - The task ID of the import image task.

$sel:architecture:ImportImage', importImageResponse_architecture - The architecture of the virtual machine.

$sel:description:ImportImage', importImageResponse_description - A description of the import task.

$sel:tags:ImportImageResponse', importImageResponse_tags - Any tags assigned to the import image task.

$sel:httpStatus:ImportImageResponse', importImageResponse_httpStatus - The response's http status code.

Response Lenses

importImageResponse_status :: Lens' ImportImageResponse (Maybe Text)

A brief status of the task.

importImageResponse_hypervisor :: Lens' ImportImageResponse (Maybe Text)

The target hypervisor of the import task.

importImageResponse_platform :: Lens' ImportImageResponse (Maybe Text)

The operating system of the virtual machine.

importImageResponse_progress :: Lens' ImportImageResponse (Maybe Text)

The progress of the task.

importImageResponse_licenseSpecifications :: Lens' ImportImageResponse (Maybe [ImportImageLicenseConfigurationResponse])

The ARNs of the license configurations.

importImageResponse_usageOperation :: Lens' ImportImageResponse (Maybe Text)

The usage operation value.

importImageResponse_licenseType :: Lens' ImportImageResponse (Maybe Text)

The license type of the virtual machine.

importImageResponse_snapshotDetails :: Lens' ImportImageResponse (Maybe [SnapshotDetail])

Information about the snapshots.

importImageResponse_encrypted :: Lens' ImportImageResponse (Maybe Bool)

Indicates whether the AMI is encrypted.

importImageResponse_kmsKeyId :: Lens' ImportImageResponse (Maybe Text)

The identifier for the symmetric KMS key that was used to create the encrypted AMI.

importImageResponse_statusMessage :: Lens' ImportImageResponse (Maybe Text)

A detailed status message of the import task.

importImageResponse_imageId :: Lens' ImportImageResponse (Maybe Text)

The ID of the Amazon Machine Image (AMI) created by the import task.

importImageResponse_importTaskId :: Lens' ImportImageResponse (Maybe Text)

The task ID of the import image task.

importImageResponse_architecture :: Lens' ImportImageResponse (Maybe Text)

The architecture of the virtual machine.

importImageResponse_description :: Lens' ImportImageResponse (Maybe Text)

A description of the import task.

importImageResponse_tags :: Lens' ImportImageResponse (Maybe [Tag])

Any tags assigned to the import image task.

importImageResponse_httpStatus :: Lens' ImportImageResponse Int

The response's http status code.