Amazonka.EC2.ImportInstance

Description

Creates an import instance task using metadata from the specified disk image.

This API action supports only single-volume VMs. To import multi-volume VMs, use ImportImage instead.

This API action is not supported by the Command Line Interface (CLI). For information about using the Amazon EC2 CLI, which is deprecated, see Importing a VM to Amazon EC2 in the Amazon EC2 CLI Reference PDF file.

For information about the import manifest referenced by this API action, see VM Import Manifest.

Creating a Request

data ImportInstance

See: newImportInstance smart constructor.

Constructors

ImportInstance'

Fields

launchSpecification :: Maybe ImportInstanceLaunchSpecification

The launch specification.

diskImages :: Maybe [DiskImage]

The disk image.

description :: Maybe Text

A description for the instance being imported.

dryRun :: Maybe Bool

Checks whether you have the required permissions for the action, without actually making the request, and provides an error response. If you have the required permissions, the error response is DryRunOperation. Otherwise, it is UnauthorizedOperation.

platform :: PlatformValues

The instance operating system.

newImportInstance

Arguments

:: PlatformValues
$sel:platform:ImportInstance'

-> ImportInstance

Create a value of ImportInstance with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:launchSpecification:ImportInstance', importInstance_launchSpecification - The launch specification.

$sel:diskImages:ImportInstance', importInstance_diskImages - The disk image.

$sel:description:ImportInstance', importInstance_description - A description for the instance being imported.

$sel:dryRun:ImportInstance', importInstance_dryRun - Checks whether you have the required permissions for the action, without actually making the request, and provides an error response. If you have the required permissions, the error response is DryRunOperation. Otherwise, it is UnauthorizedOperation.

$sel:platform:ImportInstance', importInstance_platform - The instance operating system.

Request Lenses

importInstance_launchSpecification :: Lens' ImportInstance (Maybe ImportInstanceLaunchSpecification)

The launch specification.

importInstance_diskImages :: Lens' ImportInstance (Maybe [DiskImage])

The disk image.

importInstance_description :: Lens' ImportInstance (Maybe Text)

A description for the instance being imported.

importInstance_dryRun :: Lens' ImportInstance (Maybe Bool)

Checks whether you have the required permissions for the action, without actually making the request, and provides an error response. If you have the required permissions, the error response is DryRunOperation. Otherwise, it is UnauthorizedOperation.

importInstance_platform :: Lens' ImportInstance PlatformValues

The instance operating system.

Destructuring the Response

data ImportInstanceResponse

See: newImportInstanceResponse smart constructor.

Constructors

ImportInstanceResponse'

Fields

conversionTask :: Maybe ConversionTask

Information about the conversion task.

httpStatus :: Int

The response's http status code.

newImportInstanceResponse

Arguments

:: Int
$sel:httpStatus:ImportInstanceResponse'

-> ImportInstanceResponse

Create a value of ImportInstanceResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:conversionTask:ImportInstanceResponse', importInstanceResponse_conversionTask - Information about the conversion task.

$sel:httpStatus:ImportInstanceResponse', importInstanceResponse_httpStatus - The response's http status code.

Response Lenses

importInstanceResponse_conversionTask :: Lens' ImportInstanceResponse (Maybe ConversionTask)

Information about the conversion task.

importInstanceResponse_httpStatus :: Lens' ImportInstanceResponse Int

The response's http status code.