Amazonka.EC2.ImportSnapshot

Description

Imports a disk into an EBS snapshot.

For more information, see Importing a disk as a snapshot using VM Import/Export in the VM Import/Export User Guide.

Creating a Request

data ImportSnapshot

See: newImportSnapshot smart constructor.

Constructors

ImportSnapshot'

Fields

diskContainer :: Maybe SnapshotDiskContainer

Information about the disk container.

clientToken :: Maybe Text

Token to enable idempotency for VM import requests.

roleName :: Maybe Text

The name of the role to use when not using the default role, 'vmimport'.

encrypted :: Maybe Bool

Specifies whether the destination snapshot of the imported image should be encrypted. The default KMS key for EBS is used unless you specify a non-default KMS key using KmsKeyId. For more information, see Amazon EBS Encryption in the Amazon Elastic Compute Cloud User Guide.

tagSpecifications :: Maybe [TagSpecification]

The tags to apply to the import snapshot task during creation.

kmsKeyId :: Maybe Text

An identifier for the symmetric KMS key to use when creating the encrypted snapshot. This parameter is only required if you want to use a non-default KMS key; if this parameter is not specified, the default KMS key for EBS is used. If a KmsKeyId is specified, the Encrypted flag must also be set.

The KMS key identifier may be provided in any of the following formats:

Key ID

Key alias. The alias ARN contains the arn:aws:kms namespace, followed by the Region of the key, the Amazon Web Services account ID of the key owner, the alias namespace, and then the key alias. For example, arn:aws:kms:us-east-1:012345678910:alias/ExampleAlias.

ARN using key ID. The ID ARN contains the arn:aws:kms namespace, followed by the Region of the key, the Amazon Web Services account ID of the key owner, the key namespace, and then the key ID. For example, arn:aws:kms:us-east-1:012345678910:key/abcd1234-a123-456a-a12b-a123b4cd56ef.

ARN using key alias. The alias ARN contains the arn:aws:kms namespace, followed by the Region of the key, the Amazon Web Services account ID of the key owner, the alias namespace, and then the key alias. For example, arn:aws:kms:us-east-1:012345678910:alias/ExampleAlias.

Amazon Web Services parses KmsKeyId asynchronously, meaning that the action you call may appear to complete even though you provided an invalid identifier. This action will eventually report failure.

The specified KMS key must exist in the Region that the snapshot is being copied to.

Amazon EBS does not support asymmetric KMS keys.

description :: Maybe Text

The description string for the import snapshot task.

dryRun :: Maybe Bool

Checks whether you have the required permissions for the action, without actually making the request, and provides an error response. If you have the required permissions, the error response is DryRunOperation. Otherwise, it is UnauthorizedOperation.

clientData :: Maybe ClientData

The client-specific data.

newImportSnapshot :: ImportSnapshot

Create a value of ImportSnapshot with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:diskContainer:ImportSnapshot', importSnapshot_diskContainer - Information about the disk container.

$sel:clientToken:ImportSnapshot', importSnapshot_clientToken - Token to enable idempotency for VM import requests.

$sel:roleName:ImportSnapshot', importSnapshot_roleName - The name of the role to use when not using the default role, 'vmimport'.

$sel:encrypted:ImportSnapshot', importSnapshot_encrypted - Specifies whether the destination snapshot of the imported image should be encrypted. The default KMS key for EBS is used unless you specify a non-default KMS key using KmsKeyId. For more information, see Amazon EBS Encryption in the Amazon Elastic Compute Cloud User Guide.

$sel:tagSpecifications:ImportSnapshot', importSnapshot_tagSpecifications - The tags to apply to the import snapshot task during creation.

$sel:kmsKeyId:ImportSnapshot', importSnapshot_kmsKeyId - An identifier for the symmetric KMS key to use when creating the encrypted snapshot. This parameter is only required if you want to use a non-default KMS key; if this parameter is not specified, the default KMS key for EBS is used. If a KmsKeyId is specified, the Encrypted flag must also be set.

The KMS key identifier may be provided in any of the following formats:

Key ID

Key alias. The alias ARN contains the arn:aws:kms namespace, followed by the Region of the key, the Amazon Web Services account ID of the key owner, the alias namespace, and then the key alias. For example, arn:aws:kms:us-east-1:012345678910:alias/ExampleAlias.

ARN using key ID. The ID ARN contains the arn:aws:kms namespace, followed by the Region of the key, the Amazon Web Services account ID of the key owner, the key namespace, and then the key ID. For example, arn:aws:kms:us-east-1:012345678910:key/abcd1234-a123-456a-a12b-a123b4cd56ef.

ARN using key alias. The alias ARN contains the arn:aws:kms namespace, followed by the Region of the key, the Amazon Web Services account ID of the key owner, the alias namespace, and then the key alias. For example, arn:aws:kms:us-east-1:012345678910:alias/ExampleAlias.

Amazon Web Services parses KmsKeyId asynchronously, meaning that the action you call may appear to complete even though you provided an invalid identifier. This action will eventually report failure.

The specified KMS key must exist in the Region that the snapshot is being copied to.

Amazon EBS does not support asymmetric KMS keys.

$sel:description:ImportSnapshot', importSnapshot_description - The description string for the import snapshot task.

$sel:dryRun:ImportSnapshot', importSnapshot_dryRun - Checks whether you have the required permissions for the action, without actually making the request, and provides an error response. If you have the required permissions, the error response is DryRunOperation. Otherwise, it is UnauthorizedOperation.

$sel:clientData:ImportSnapshot', importSnapshot_clientData - The client-specific data.

Request Lenses

importSnapshot_diskContainer :: Lens' ImportSnapshot (Maybe SnapshotDiskContainer)

Information about the disk container.

importSnapshot_clientToken :: Lens' ImportSnapshot (Maybe Text)

Token to enable idempotency for VM import requests.

importSnapshot_roleName :: Lens' ImportSnapshot (Maybe Text)

The name of the role to use when not using the default role, 'vmimport'.

importSnapshot_encrypted :: Lens' ImportSnapshot (Maybe Bool)

Specifies whether the destination snapshot of the imported image should be encrypted. The default KMS key for EBS is used unless you specify a non-default KMS key using KmsKeyId. For more information, see Amazon EBS Encryption in the Amazon Elastic Compute Cloud User Guide.

importSnapshot_tagSpecifications :: Lens' ImportSnapshot (Maybe [TagSpecification])

The tags to apply to the import snapshot task during creation.

importSnapshot_kmsKeyId :: Lens' ImportSnapshot (Maybe Text)

An identifier for the symmetric KMS key to use when creating the encrypted snapshot. This parameter is only required if you want to use a non-default KMS key; if this parameter is not specified, the default KMS key for EBS is used. If a KmsKeyId is specified, the Encrypted flag must also be set.

The KMS key identifier may be provided in any of the following formats:

Key ID

Key alias. The alias ARN contains the arn:aws:kms namespace, followed by the Region of the key, the Amazon Web Services account ID of the key owner, the alias namespace, and then the key alias. For example, arn:aws:kms:us-east-1:012345678910:alias/ExampleAlias.

ARN using key ID. The ID ARN contains the arn:aws:kms namespace, followed by the Region of the key, the Amazon Web Services account ID of the key owner, the key namespace, and then the key ID. For example, arn:aws:kms:us-east-1:012345678910:key/abcd1234-a123-456a-a12b-a123b4cd56ef.

ARN using key alias. The alias ARN contains the arn:aws:kms namespace, followed by the Region of the key, the Amazon Web Services account ID of the key owner, the alias namespace, and then the key alias. For example, arn:aws:kms:us-east-1:012345678910:alias/ExampleAlias.

Amazon Web Services parses KmsKeyId asynchronously, meaning that the action you call may appear to complete even though you provided an invalid identifier. This action will eventually report failure.

The specified KMS key must exist in the Region that the snapshot is being copied to.

Amazon EBS does not support asymmetric KMS keys.

importSnapshot_description :: Lens' ImportSnapshot (Maybe Text)

The description string for the import snapshot task.

importSnapshot_dryRun :: Lens' ImportSnapshot (Maybe Bool)

Checks whether you have the required permissions for the action, without actually making the request, and provides an error response. If you have the required permissions, the error response is DryRunOperation. Otherwise, it is UnauthorizedOperation.

importSnapshot_clientData :: Lens' ImportSnapshot (Maybe ClientData)

The client-specific data.

Destructuring the Response

data ImportSnapshotResponse

See: newImportSnapshotResponse smart constructor.

Constructors

ImportSnapshotResponse'

Fields

snapshotTaskDetail :: Maybe SnapshotTaskDetail

Information about the import snapshot task.

importTaskId :: Maybe Text

The ID of the import snapshot task.

description :: Maybe Text

A description of the import snapshot task.

tags :: Maybe [Tag]

Any tags assigned to the import snapshot task.

httpStatus :: Int

The response's http status code.

newImportSnapshotResponse

Arguments

:: Int
$sel:httpStatus:ImportSnapshotResponse'

-> ImportSnapshotResponse

Create a value of ImportSnapshotResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:snapshotTaskDetail:ImportSnapshotResponse', importSnapshotResponse_snapshotTaskDetail - Information about the import snapshot task.

$sel:importTaskId:ImportSnapshotResponse', importSnapshotResponse_importTaskId - The ID of the import snapshot task.

$sel:description:ImportSnapshot', importSnapshotResponse_description - A description of the import snapshot task.

$sel:tags:ImportSnapshotResponse', importSnapshotResponse_tags - Any tags assigned to the import snapshot task.

$sel:httpStatus:ImportSnapshotResponse', importSnapshotResponse_httpStatus - The response's http status code.

Response Lenses

importSnapshotResponse_snapshotTaskDetail :: Lens' ImportSnapshotResponse (Maybe SnapshotTaskDetail)

Information about the import snapshot task.

importSnapshotResponse_importTaskId :: Lens' ImportSnapshotResponse (Maybe Text)

The ID of the import snapshot task.

importSnapshotResponse_description :: Lens' ImportSnapshotResponse (Maybe Text)

A description of the import snapshot task.

importSnapshotResponse_tags :: Lens' ImportSnapshotResponse (Maybe [Tag])

Any tags assigned to the import snapshot task.

importSnapshotResponse_httpStatus :: Lens' ImportSnapshotResponse Int

The response's http status code.