Amazonka.EC2.ModifyAddressAttribute

Description

Modifies an attribute of the specified Elastic IP address. For requirements, see Using reverse DNS for email applications.

Creating a Request

data ModifyAddressAttribute

See: newModifyAddressAttribute smart constructor.

Constructors

ModifyAddressAttribute'

Fields

domainName :: Maybe Text

The domain name to modify for the IP address.

dryRun :: Maybe Bool

Checks whether you have the required permissions for the action, without actually making the request, and provides an error response. If you have the required permissions, the error response is DryRunOperation. Otherwise, it is UnauthorizedOperation.

allocationId :: Text

EC2-VPC

The allocation ID.

newModifyAddressAttribute

Arguments

:: Text
$sel:allocationId:ModifyAddressAttribute'

-> ModifyAddressAttribute

Create a value of ModifyAddressAttribute with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:domainName:ModifyAddressAttribute', modifyAddressAttribute_domainName - The domain name to modify for the IP address.

$sel:dryRun:ModifyAddressAttribute', modifyAddressAttribute_dryRun - Checks whether you have the required permissions for the action, without actually making the request, and provides an error response. If you have the required permissions, the error response is DryRunOperation. Otherwise, it is UnauthorizedOperation.

$sel:allocationId:ModifyAddressAttribute', modifyAddressAttribute_allocationId - [EC2-VPC] The allocation ID.

Request Lenses

modifyAddressAttribute_domainName :: Lens' ModifyAddressAttribute (Maybe Text)

The domain name to modify for the IP address.

modifyAddressAttribute_dryRun :: Lens' ModifyAddressAttribute (Maybe Bool)

Checks whether you have the required permissions for the action, without actually making the request, and provides an error response. If you have the required permissions, the error response is DryRunOperation. Otherwise, it is UnauthorizedOperation.

modifyAddressAttribute_allocationId :: Lens' ModifyAddressAttribute Text

EC2-VPC

The allocation ID.

Destructuring the Response

data ModifyAddressAttributeResponse

See: newModifyAddressAttributeResponse smart constructor.

Constructors

ModifyAddressAttributeResponse'

Fields

address :: Maybe AddressAttribute

Information about the Elastic IP address.

httpStatus :: Int

The response's http status code.

newModifyAddressAttributeResponse

Arguments

:: Int
$sel:httpStatus:ModifyAddressAttributeResponse'

-> ModifyAddressAttributeResponse

Create a value of ModifyAddressAttributeResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:address:ModifyAddressAttributeResponse', modifyAddressAttributeResponse_address - Information about the Elastic IP address.

$sel:httpStatus:ModifyAddressAttributeResponse', modifyAddressAttributeResponse_httpStatus - The response's http status code.

Response Lenses

modifyAddressAttributeResponse_address :: Lens' ModifyAddressAttributeResponse (Maybe AddressAttribute)

Information about the Elastic IP address.

modifyAddressAttributeResponse_httpStatus :: Lens' ModifyAddressAttributeResponse Int

The response's http status code.