Amazonka.EC2.ModifyImageAttribute

Description

Modifies the specified attribute of the specified AMI. You can specify only one attribute at a time. You can use the Attribute parameter to specify the attribute or one of the following parameters: Description or LaunchPermission.

Images with an Amazon Web Services Marketplace product code cannot be made public.

To enable the SriovNetSupport enhanced networking attribute of an image, enable SriovNetSupport on an instance and create an AMI from the instance.

Creating a Request

data ModifyImageAttribute

Contains the parameters for ModifyImageAttribute.

See: newModifyImageAttribute smart constructor.

Constructors

ModifyImageAttribute'

Fields

attribute :: Maybe Text

The name of the attribute to modify. The valid values are description and launchPermission.

userIds :: Maybe [Text]

The Amazon Web Services account IDs. This parameter can be used only when the Attribute parameter is launchPermission.

userGroups :: Maybe [Text]

The user groups. This parameter can be used only when the Attribute parameter is launchPermission.

value :: Maybe Text

The value of the attribute being modified. This parameter can be used only when the Attribute parameter is description.

launchPermission :: Maybe LaunchPermissionModifications

A new launch permission for the AMI.

operationType :: Maybe OperationType

The operation type. This parameter can be used only when the Attribute parameter is launchPermission.

productCodes :: Maybe [Text]

Not supported.

description :: Maybe AttributeValue

A new description for the AMI.

dryRun :: Maybe Bool

Checks whether you have the required permissions for the action, without actually making the request, and provides an error response. If you have the required permissions, the error response is DryRunOperation. Otherwise, it is UnauthorizedOperation.

imageId :: Text

The ID of the AMI.

newModifyImageAttribute

Arguments

:: Text
$sel:imageId:ModifyImageAttribute'

-> ModifyImageAttribute

Create a value of ModifyImageAttribute with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:attribute:ModifyImageAttribute', modifyImageAttribute_attribute - The name of the attribute to modify. The valid values are description and launchPermission.

$sel:userIds:ModifyImageAttribute', modifyImageAttribute_userIds - The Amazon Web Services account IDs. This parameter can be used only when the Attribute parameter is launchPermission.

$sel:userGroups:ModifyImageAttribute', modifyImageAttribute_userGroups - The user groups. This parameter can be used only when the Attribute parameter is launchPermission.

$sel:value:ModifyImageAttribute', modifyImageAttribute_value - The value of the attribute being modified. This parameter can be used only when the Attribute parameter is description.

$sel:launchPermission:ModifyImageAttribute', modifyImageAttribute_launchPermission - A new launch permission for the AMI.

$sel:operationType:ModifyImageAttribute', modifyImageAttribute_operationType - The operation type. This parameter can be used only when the Attribute parameter is launchPermission.

$sel:productCodes:ModifyImageAttribute', modifyImageAttribute_productCodes - Not supported.

$sel:description:ModifyImageAttribute', modifyImageAttribute_description - A new description for the AMI.

$sel:dryRun:ModifyImageAttribute', modifyImageAttribute_dryRun - Checks whether you have the required permissions for the action, without actually making the request, and provides an error response. If you have the required permissions, the error response is DryRunOperation. Otherwise, it is UnauthorizedOperation.

$sel:imageId:ModifyImageAttribute', modifyImageAttribute_imageId - The ID of the AMI.

Request Lenses

modifyImageAttribute_attribute :: Lens' ModifyImageAttribute (Maybe Text)

The name of the attribute to modify. The valid values are description and launchPermission.

modifyImageAttribute_userIds :: Lens' ModifyImageAttribute (Maybe [Text])

The Amazon Web Services account IDs. This parameter can be used only when the Attribute parameter is launchPermission.

modifyImageAttribute_userGroups :: Lens' ModifyImageAttribute (Maybe [Text])

The user groups. This parameter can be used only when the Attribute parameter is launchPermission.

modifyImageAttribute_value :: Lens' ModifyImageAttribute (Maybe Text)

The value of the attribute being modified. This parameter can be used only when the Attribute parameter is description.

modifyImageAttribute_launchPermission :: Lens' ModifyImageAttribute (Maybe LaunchPermissionModifications)

A new launch permission for the AMI.

modifyImageAttribute_operationType :: Lens' ModifyImageAttribute (Maybe OperationType)

The operation type. This parameter can be used only when the Attribute parameter is launchPermission.

modifyImageAttribute_productCodes :: Lens' ModifyImageAttribute (Maybe [Text])

Not supported.

modifyImageAttribute_description :: Lens' ModifyImageAttribute (Maybe AttributeValue)

A new description for the AMI.

modifyImageAttribute_dryRun :: Lens' ModifyImageAttribute (Maybe Bool)

Checks whether you have the required permissions for the action, without actually making the request, and provides an error response. If you have the required permissions, the error response is DryRunOperation. Otherwise, it is UnauthorizedOperation.

modifyImageAttribute_imageId :: Lens' ModifyImageAttribute Text

The ID of the AMI.

Destructuring the Response

data ModifyImageAttributeResponse

See: newModifyImageAttributeResponse smart constructor.

Constructors

ModifyImageAttributeResponse'

newModifyImageAttributeResponse :: ModifyImageAttributeResponse

Create a value of ModifyImageAttributeResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.