Amazonka.EC2.ModifyInstanceEventStartTime

Description

Modifies the start time for a scheduled Amazon EC2 instance event.

Creating a Request

data ModifyInstanceEventStartTime

See: newModifyInstanceEventStartTime smart constructor.

Constructors

ModifyInstanceEventStartTime'

Fields

dryRun :: Maybe Bool

Checks whether you have the required permissions for the action, without actually making the request, and provides an error response. If you have the required permissions, the error response is DryRunOperation. Otherwise, it is UnauthorizedOperation.

instanceId :: Text

The ID of the instance with the scheduled event.

instanceEventId :: Text

The ID of the event whose date and time you are modifying.

notBefore :: ISO8601

The new date and time when the event will take place.

newModifyInstanceEventStartTime

Arguments

:: Text
$sel:instanceId:ModifyInstanceEventStartTime'

-> Text
$sel:instanceEventId:ModifyInstanceEventStartTime'

-> UTCTime
$sel:notBefore:ModifyInstanceEventStartTime'

-> ModifyInstanceEventStartTime

Create a value of ModifyInstanceEventStartTime with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:dryRun:ModifyInstanceEventStartTime', modifyInstanceEventStartTime_dryRun - Checks whether you have the required permissions for the action, without actually making the request, and provides an error response. If you have the required permissions, the error response is DryRunOperation. Otherwise, it is UnauthorizedOperation.

$sel:instanceId:ModifyInstanceEventStartTime', modifyInstanceEventStartTime_instanceId - The ID of the instance with the scheduled event.

$sel:instanceEventId:ModifyInstanceEventStartTime', modifyInstanceEventStartTime_instanceEventId - The ID of the event whose date and time you are modifying.

$sel:notBefore:ModifyInstanceEventStartTime', modifyInstanceEventStartTime_notBefore - The new date and time when the event will take place.

Request Lenses

modifyInstanceEventStartTime_dryRun :: Lens' ModifyInstanceEventStartTime (Maybe Bool)

Checks whether you have the required permissions for the action, without actually making the request, and provides an error response. If you have the required permissions, the error response is DryRunOperation. Otherwise, it is UnauthorizedOperation.

modifyInstanceEventStartTime_instanceId :: Lens' ModifyInstanceEventStartTime Text

The ID of the instance with the scheduled event.

modifyInstanceEventStartTime_instanceEventId :: Lens' ModifyInstanceEventStartTime Text

The ID of the event whose date and time you are modifying.

modifyInstanceEventStartTime_notBefore :: Lens' ModifyInstanceEventStartTime UTCTime

The new date and time when the event will take place.

Destructuring the Response

data ModifyInstanceEventStartTimeResponse

See: newModifyInstanceEventStartTimeResponse smart constructor.

Constructors

ModifyInstanceEventStartTimeResponse'

Fields

event :: Maybe InstanceStatusEvent

httpStatus :: Int

The response's http status code.

newModifyInstanceEventStartTimeResponse

Arguments

:: Int
$sel:httpStatus:ModifyInstanceEventStartTimeResponse'

-> ModifyInstanceEventStartTimeResponse

Create a value of ModifyInstanceEventStartTimeResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:event:ModifyInstanceEventStartTimeResponse', modifyInstanceEventStartTimeResponse_event - Undocumented member.

$sel:httpStatus:ModifyInstanceEventStartTimeResponse', modifyInstanceEventStartTimeResponse_httpStatus - The response's http status code.

Response Lenses

modifyInstanceEventStartTimeResponse_event :: Lens' ModifyInstanceEventStartTimeResponse (Maybe InstanceStatusEvent)

Undocumented member.

modifyInstanceEventStartTimeResponse_httpStatus :: Lens' ModifyInstanceEventStartTimeResponse Int

The response's http status code.