Amazonka.EC2.ModifyInstancePlacement

Description

Modifies the placement attributes for a specified instance. You can do the following:

Modify the affinity between an instance and a Dedicated Host. When affinity is set to host and the instance is not associated with a specific Dedicated Host, the next time the instance is launched, it is automatically associated with the host on which it lands. If the instance is restarted or rebooted, this relationship persists.

Change the Dedicated Host with which an instance is associated.

Change the instance tenancy of an instance.

Move an instance to or from a placement group.

At least one attribute for affinity, host ID, tenancy, or placement group name must be specified in the request. Affinity and tenancy can be modified in the same request.

To modify the host ID, tenancy, placement group, or partition for an instance, the instance must be in the stopped state.

Creating a Request

data ModifyInstancePlacement

See: newModifyInstancePlacement smart constructor.

Constructors

ModifyInstancePlacement'

Fields

affinity :: Maybe Affinity

The affinity setting for the instance.

hostId :: Maybe Text

The ID of the Dedicated Host with which to associate the instance.

partitionNumber :: Maybe Int

Reserved for future use.

tenancy :: Maybe HostTenancy

The tenancy for the instance.

For T3 instances, you can't change the tenancy from dedicated to host, or from host to dedicated. Attempting to make one of these unsupported tenancy changes results in the InvalidTenancy error code.

groupName :: Maybe Text

The name of the placement group in which to place the instance. For spread placement groups, the instance must have a tenancy of default. For cluster and partition placement groups, the instance must have a tenancy of default or dedicated.

To remove an instance from a placement group, specify an empty string ("").

hostResourceGroupArn :: Maybe Text

The ARN of the host resource group in which to place the instance.

instanceId :: Text

The ID of the instance that you are modifying.

newModifyInstancePlacement

Arguments

:: Text
$sel:instanceId:ModifyInstancePlacement'

-> ModifyInstancePlacement

Create a value of ModifyInstancePlacement with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:affinity:ModifyInstancePlacement', modifyInstancePlacement_affinity - The affinity setting for the instance.

$sel:hostId:ModifyInstancePlacement', modifyInstancePlacement_hostId - The ID of the Dedicated Host with which to associate the instance.

$sel:partitionNumber:ModifyInstancePlacement', modifyInstancePlacement_partitionNumber - Reserved for future use.

$sel:tenancy:ModifyInstancePlacement', modifyInstancePlacement_tenancy - The tenancy for the instance.

For T3 instances, you can't change the tenancy from dedicated to host, or from host to dedicated. Attempting to make one of these unsupported tenancy changes results in the InvalidTenancy error code.

$sel:groupName:ModifyInstancePlacement', modifyInstancePlacement_groupName - The name of the placement group in which to place the instance. For spread placement groups, the instance must have a tenancy of default. For cluster and partition placement groups, the instance must have a tenancy of default or dedicated.

To remove an instance from a placement group, specify an empty string ("").

$sel:hostResourceGroupArn:ModifyInstancePlacement', modifyInstancePlacement_hostResourceGroupArn - The ARN of the host resource group in which to place the instance.

$sel:instanceId:ModifyInstancePlacement', modifyInstancePlacement_instanceId - The ID of the instance that you are modifying.

Request Lenses

modifyInstancePlacement_affinity :: Lens' ModifyInstancePlacement (Maybe Affinity)

The affinity setting for the instance.

modifyInstancePlacement_hostId :: Lens' ModifyInstancePlacement (Maybe Text)

The ID of the Dedicated Host with which to associate the instance.

modifyInstancePlacement_partitionNumber :: Lens' ModifyInstancePlacement (Maybe Int)

Reserved for future use.

modifyInstancePlacement_tenancy :: Lens' ModifyInstancePlacement (Maybe HostTenancy)

The tenancy for the instance.

For T3 instances, you can't change the tenancy from dedicated to host, or from host to dedicated. Attempting to make one of these unsupported tenancy changes results in the InvalidTenancy error code.

modifyInstancePlacement_groupName :: Lens' ModifyInstancePlacement (Maybe Text)

The name of the placement group in which to place the instance. For spread placement groups, the instance must have a tenancy of default. For cluster and partition placement groups, the instance must have a tenancy of default or dedicated.

To remove an instance from a placement group, specify an empty string ("").

modifyInstancePlacement_hostResourceGroupArn :: Lens' ModifyInstancePlacement (Maybe Text)

The ARN of the host resource group in which to place the instance.

modifyInstancePlacement_instanceId :: Lens' ModifyInstancePlacement Text

The ID of the instance that you are modifying.

Destructuring the Response

data ModifyInstancePlacementResponse

See: newModifyInstancePlacementResponse smart constructor.

Constructors

ModifyInstancePlacementResponse'

Fields

return' :: Maybe Bool

Is true if the request succeeds, and an error otherwise.

httpStatus :: Int

The response's http status code.

newModifyInstancePlacementResponse

Arguments

:: Int
$sel:httpStatus:ModifyInstancePlacementResponse'

-> ModifyInstancePlacementResponse

Create a value of ModifyInstancePlacementResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:return':ModifyInstancePlacementResponse', modifyInstancePlacementResponse_return - Is true if the request succeeds, and an error otherwise.

$sel:httpStatus:ModifyInstancePlacementResponse', modifyInstancePlacementResponse_httpStatus - The response's http status code.

Response Lenses

modifyInstancePlacementResponse_return :: Lens' ModifyInstancePlacementResponse (Maybe Bool)

Is true if the request succeeds, and an error otherwise.

modifyInstancePlacementResponse_httpStatus :: Lens' ModifyInstancePlacementResponse Int

The response's http status code.