Amazonka.EC2.ModifySnapshotAttribute

Description

Adds or removes permission settings for the specified snapshot. You may add or remove specified Amazon Web Services account IDs from a snapshot's list of create volume permissions, but you cannot do both in a single operation. If you need to both add and remove account IDs for a snapshot, you must use multiple operations. You can make up to 500 modifications to a snapshot in a single operation.

Encrypted snapshots and snapshots with Amazon Web Services Marketplace product codes cannot be made public. Snapshots encrypted with your default KMS key cannot be shared with other accounts.

For more information about modifying snapshot permissions, see Share a snapshot in the Amazon Elastic Compute Cloud User Guide.

Creating a Request

data ModifySnapshotAttribute

See: newModifySnapshotAttribute smart constructor.

Constructors

ModifySnapshotAttribute'

Fields

attribute :: Maybe SnapshotAttributeName

The snapshot attribute to modify. Only volume creation permissions can be modified.

createVolumePermission :: Maybe CreateVolumePermissionModifications

A JSON representation of the snapshot attribute modification.

userIds :: Maybe [Text]

The account ID to modify for the snapshot.

groupNames :: Maybe [Text]

The group to modify for the snapshot.

operationType :: Maybe OperationType

The type of operation to perform to the attribute.

dryRun :: Maybe Bool

Checks whether you have the required permissions for the action, without actually making the request, and provides an error response. If you have the required permissions, the error response is DryRunOperation. Otherwise, it is UnauthorizedOperation.

snapshotId :: Text

The ID of the snapshot.

newModifySnapshotAttribute

Arguments

:: Text
$sel:snapshotId:ModifySnapshotAttribute'

-> ModifySnapshotAttribute

Create a value of ModifySnapshotAttribute with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:attribute:ModifySnapshotAttribute', modifySnapshotAttribute_attribute - The snapshot attribute to modify. Only volume creation permissions can be modified.

$sel:createVolumePermission:ModifySnapshotAttribute', modifySnapshotAttribute_createVolumePermission - A JSON representation of the snapshot attribute modification.

$sel:userIds:ModifySnapshotAttribute', modifySnapshotAttribute_userIds - The account ID to modify for the snapshot.

$sel:groupNames:ModifySnapshotAttribute', modifySnapshotAttribute_groupNames - The group to modify for the snapshot.

$sel:operationType:ModifySnapshotAttribute', modifySnapshotAttribute_operationType - The type of operation to perform to the attribute.

$sel:dryRun:ModifySnapshotAttribute', modifySnapshotAttribute_dryRun - Checks whether you have the required permissions for the action, without actually making the request, and provides an error response. If you have the required permissions, the error response is DryRunOperation. Otherwise, it is UnauthorizedOperation.

$sel:snapshotId:ModifySnapshotAttribute', modifySnapshotAttribute_snapshotId - The ID of the snapshot.

Request Lenses

modifySnapshotAttribute_attribute :: Lens' ModifySnapshotAttribute (Maybe SnapshotAttributeName)

The snapshot attribute to modify. Only volume creation permissions can be modified.

modifySnapshotAttribute_createVolumePermission :: Lens' ModifySnapshotAttribute (Maybe CreateVolumePermissionModifications)

A JSON representation of the snapshot attribute modification.

modifySnapshotAttribute_userIds :: Lens' ModifySnapshotAttribute (Maybe [Text])

The account ID to modify for the snapshot.

modifySnapshotAttribute_groupNames :: Lens' ModifySnapshotAttribute (Maybe [Text])

The group to modify for the snapshot.

modifySnapshotAttribute_operationType :: Lens' ModifySnapshotAttribute (Maybe OperationType)

The type of operation to perform to the attribute.

modifySnapshotAttribute_dryRun :: Lens' ModifySnapshotAttribute (Maybe Bool)

Checks whether you have the required permissions for the action, without actually making the request, and provides an error response. If you have the required permissions, the error response is DryRunOperation. Otherwise, it is UnauthorizedOperation.

modifySnapshotAttribute_snapshotId :: Lens' ModifySnapshotAttribute Text

The ID of the snapshot.

Destructuring the Response

data ModifySnapshotAttributeResponse

See: newModifySnapshotAttributeResponse smart constructor.

Constructors

ModifySnapshotAttributeResponse'

newModifySnapshotAttributeResponse :: ModifySnapshotAttributeResponse

Create a value of ModifySnapshotAttributeResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.