Amazonka.EC2.ModifyVolume

Description

You can modify several parameters of an existing EBS volume, including volume size, volume type, and IOPS capacity. If your EBS volume is attached to a current-generation EC2 instance type, you might be able to apply these changes without stopping the instance or detaching the volume from it. For more information about modifying EBS volumes, see Amazon EBS Elastic Volumes (Linux instances) or Amazon EBS Elastic Volumes (Windows instances).

When you complete a resize operation on your volume, you need to extend the volume's file-system size to take advantage of the new storage capacity. For more information, see Extend a Linux file system or Extend a Windows file system.

You can use CloudWatch Events to check the status of a modification to an EBS volume. For information about CloudWatch Events, see the Amazon CloudWatch Events User Guide. You can also track the status of a modification using DescribeVolumesModifications. For information about tracking status changes using either method, see Monitor the progress of volume modifications.

With previous-generation instance types, resizing an EBS volume might require detaching and reattaching the volume or stopping and restarting the instance.

If you reach the maximum volume modification rate per volume limit, you must wait at least six hours before applying further modifications to the affected EBS volume.

Creating a Request

data ModifyVolume

See: newModifyVolume smart constructor.

Constructors

ModifyVolume'

Fields

multiAttachEnabled :: Maybe Bool

Specifies whether to enable Amazon EBS Multi-Attach. If you enable Multi-Attach, you can attach the volume to up to 16 Nitro-based instances in the same Availability Zone. This parameter is supported with io1 and io2 volumes only. For more information, see Amazon EBS Multi-Attach in the Amazon Elastic Compute Cloud User Guide.

throughput :: Maybe Int

The target throughput of the volume, in MiB/s. This parameter is valid only for gp3 volumes. The maximum value is 1,000.

Default: The existing value is retained if the source and target volume type is gp3. Otherwise, the default value is 125.

Valid Range: Minimum value of 125. Maximum value of 1000.

size :: Maybe Int

The target size of the volume, in GiB. The target volume size must be greater than or equal to the existing size of the volume.

The following are the supported volumes sizes for each volume type:

gp2 and gp3: 1-16,384

io1 and io2: 4-16,384

st1 and sc1: 125-16,384

standard: 1-1,024

Default: The existing size is retained.

iops :: Maybe Int

The target IOPS rate of the volume. This parameter is valid only for gp3, io1, and io2 volumes.

The following are the supported values for each volume type:

gp3: 3,000-16,000 IOPS

io1: 100-64,000 IOPS

io2: 100-64,000 IOPS

Default: The existing value is retained if you keep the same volume type. If you change the volume type to io1, io2, or gp3, the default is 3,000.

volumeType :: Maybe VolumeType

The target EBS volume type of the volume. For more information, see Amazon EBS volume types in the Amazon Elastic Compute Cloud User Guide.

Default: The existing type is retained.

dryRun :: Maybe Bool

Checks whether you have the required permissions for the action, without actually making the request, and provides an error response. If you have the required permissions, the error response is DryRunOperation. Otherwise, it is UnauthorizedOperation.

volumeId :: Text

The ID of the volume.

newModifyVolume

Arguments

:: Text
$sel:volumeId:ModifyVolume'

-> ModifyVolume

Create a value of ModifyVolume with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:multiAttachEnabled:ModifyVolume', modifyVolume_multiAttachEnabled - Specifies whether to enable Amazon EBS Multi-Attach. If you enable Multi-Attach, you can attach the volume to up to 16 Nitro-based instances in the same Availability Zone. This parameter is supported with io1 and io2 volumes only. For more information, see Amazon EBS Multi-Attach in the Amazon Elastic Compute Cloud User Guide.

$sel:throughput:ModifyVolume', modifyVolume_throughput - The target throughput of the volume, in MiB/s. This parameter is valid only for gp3 volumes. The maximum value is 1,000.

Default: The existing value is retained if the source and target volume type is gp3. Otherwise, the default value is 125.

Valid Range: Minimum value of 125. Maximum value of 1000.

$sel:size:ModifyVolume', modifyVolume_size - The target size of the volume, in GiB. The target volume size must be greater than or equal to the existing size of the volume.

The following are the supported volumes sizes for each volume type:

gp2 and gp3: 1-16,384

io1 and io2: 4-16,384

st1 and sc1: 125-16,384

standard: 1-1,024

Default: The existing size is retained.

$sel:iops:ModifyVolume', modifyVolume_iops - The target IOPS rate of the volume. This parameter is valid only for gp3, io1, and io2 volumes.

The following are the supported values for each volume type:

gp3: 3,000-16,000 IOPS

io1: 100-64,000 IOPS

io2: 100-64,000 IOPS

Default: The existing value is retained if you keep the same volume type. If you change the volume type to io1, io2, or gp3, the default is 3,000.

$sel:volumeType:ModifyVolume', modifyVolume_volumeType - The target EBS volume type of the volume. For more information, see Amazon EBS volume types in the Amazon Elastic Compute Cloud User Guide.

Default: The existing type is retained.

$sel:dryRun:ModifyVolume', modifyVolume_dryRun - Checks whether you have the required permissions for the action, without actually making the request, and provides an error response. If you have the required permissions, the error response is DryRunOperation. Otherwise, it is UnauthorizedOperation.

$sel:volumeId:ModifyVolume', modifyVolume_volumeId - The ID of the volume.

Request Lenses

modifyVolume_multiAttachEnabled :: Lens' ModifyVolume (Maybe Bool)

Specifies whether to enable Amazon EBS Multi-Attach. If you enable Multi-Attach, you can attach the volume to up to 16 Nitro-based instances in the same Availability Zone. This parameter is supported with io1 and io2 volumes only. For more information, see Amazon EBS Multi-Attach in the Amazon Elastic Compute Cloud User Guide.

modifyVolume_throughput :: Lens' ModifyVolume (Maybe Int)

The target throughput of the volume, in MiB/s. This parameter is valid only for gp3 volumes. The maximum value is 1,000.

Default: The existing value is retained if the source and target volume type is gp3. Otherwise, the default value is 125.

Valid Range: Minimum value of 125. Maximum value of 1000.

modifyVolume_size :: Lens' ModifyVolume (Maybe Int)

The target size of the volume, in GiB. The target volume size must be greater than or equal to the existing size of the volume.

The following are the supported volumes sizes for each volume type:

gp2 and gp3: 1-16,384

io1 and io2: 4-16,384

st1 and sc1: 125-16,384

standard: 1-1,024

Default: The existing size is retained.

modifyVolume_iops :: Lens' ModifyVolume (Maybe Int)

The target IOPS rate of the volume. This parameter is valid only for gp3, io1, and io2 volumes.

The following are the supported values for each volume type:

gp3: 3,000-16,000 IOPS

io1: 100-64,000 IOPS

io2: 100-64,000 IOPS

Default: The existing value is retained if you keep the same volume type. If you change the volume type to io1, io2, or gp3, the default is 3,000.

modifyVolume_volumeType :: Lens' ModifyVolume (Maybe VolumeType)

The target EBS volume type of the volume. For more information, see Amazon EBS volume types in the Amazon Elastic Compute Cloud User Guide.

Default: The existing type is retained.

modifyVolume_dryRun :: Lens' ModifyVolume (Maybe Bool)

Checks whether you have the required permissions for the action, without actually making the request, and provides an error response. If you have the required permissions, the error response is DryRunOperation. Otherwise, it is UnauthorizedOperation.

modifyVolume_volumeId :: Lens' ModifyVolume Text

The ID of the volume.

Destructuring the Response

data ModifyVolumeResponse

See: newModifyVolumeResponse smart constructor.

Constructors

ModifyVolumeResponse'

Fields

volumeModification :: Maybe VolumeModification

Information about the volume modification.

httpStatus :: Int

The response's http status code.

newModifyVolumeResponse

Arguments

:: Int
$sel:httpStatus:ModifyVolumeResponse'

-> ModifyVolumeResponse

Create a value of ModifyVolumeResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:volumeModification:ModifyVolumeResponse', modifyVolumeResponse_volumeModification - Information about the volume modification.

$sel:httpStatus:ModifyVolumeResponse', modifyVolumeResponse_httpStatus - The response's http status code.

Response Lenses

modifyVolumeResponse_volumeModification :: Lens' ModifyVolumeResponse (Maybe VolumeModification)

Information about the volume modification.

modifyVolumeResponse_httpStatus :: Lens' ModifyVolumeResponse Int

The response's http status code.