Amazonka.EC2.ModifyVpnConnectionOptions

Description

Modifies the connection options for your Site-to-Site VPN connection.

When you modify the VPN connection options, the VPN endpoint IP addresses on the Amazon Web Services side do not change, and the tunnel options do not change. Your VPN connection will be temporarily unavailable for a brief period while the VPN connection is updated.

Creating a Request

data ModifyVpnConnectionOptions

See: newModifyVpnConnectionOptions smart constructor.

Constructors

ModifyVpnConnectionOptions'

Fields

remoteIpv4NetworkCidr :: Maybe Text

The IPv4 CIDR on the Amazon Web Services side of the VPN connection.

Default: 0.0.0.0/0

localIpv4NetworkCidr :: Maybe Text

The IPv4 CIDR on the customer gateway (on-premises) side of the VPN connection.

Default: 0.0.0.0/0

remoteIpv6NetworkCidr :: Maybe Text

The IPv6 CIDR on the Amazon Web Services side of the VPN connection.

Default: ::/0

localIpv6NetworkCidr :: Maybe Text

The IPv6 CIDR on the customer gateway (on-premises) side of the VPN connection.

Default: ::/0

dryRun :: Maybe Bool

Checks whether you have the required permissions for the action, without actually making the request, and provides an error response. If you have the required permissions, the error response is DryRunOperation. Otherwise, it is UnauthorizedOperation.

vpnConnectionId :: Text

The ID of the Site-to-Site VPN connection.

newModifyVpnConnectionOptions

Arguments

:: Text
$sel:vpnConnectionId:ModifyVpnConnectionOptions'

-> ModifyVpnConnectionOptions

Create a value of ModifyVpnConnectionOptions with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:remoteIpv4NetworkCidr:ModifyVpnConnectionOptions', modifyVpnConnectionOptions_remoteIpv4NetworkCidr - The IPv4 CIDR on the Amazon Web Services side of the VPN connection.

Default: 0.0.0.0/0

$sel:localIpv4NetworkCidr:ModifyVpnConnectionOptions', modifyVpnConnectionOptions_localIpv4NetworkCidr - The IPv4 CIDR on the customer gateway (on-premises) side of the VPN connection.

Default: 0.0.0.0/0

$sel:remoteIpv6NetworkCidr:ModifyVpnConnectionOptions', modifyVpnConnectionOptions_remoteIpv6NetworkCidr - The IPv6 CIDR on the Amazon Web Services side of the VPN connection.

Default: ::/0

$sel:localIpv6NetworkCidr:ModifyVpnConnectionOptions', modifyVpnConnectionOptions_localIpv6NetworkCidr - The IPv6 CIDR on the customer gateway (on-premises) side of the VPN connection.

Default: ::/0

$sel:dryRun:ModifyVpnConnectionOptions', modifyVpnConnectionOptions_dryRun - Checks whether you have the required permissions for the action, without actually making the request, and provides an error response. If you have the required permissions, the error response is DryRunOperation. Otherwise, it is UnauthorizedOperation.

$sel:vpnConnectionId:ModifyVpnConnectionOptions', modifyVpnConnectionOptions_vpnConnectionId - The ID of the Site-to-Site VPN connection.

Request Lenses

modifyVpnConnectionOptions_remoteIpv4NetworkCidr :: Lens' ModifyVpnConnectionOptions (Maybe Text)

The IPv4 CIDR on the Amazon Web Services side of the VPN connection.

Default: 0.0.0.0/0

modifyVpnConnectionOptions_localIpv4NetworkCidr :: Lens' ModifyVpnConnectionOptions (Maybe Text)

The IPv4 CIDR on the customer gateway (on-premises) side of the VPN connection.

Default: 0.0.0.0/0

modifyVpnConnectionOptions_remoteIpv6NetworkCidr :: Lens' ModifyVpnConnectionOptions (Maybe Text)

The IPv6 CIDR on the Amazon Web Services side of the VPN connection.

Default: ::/0

modifyVpnConnectionOptions_localIpv6NetworkCidr :: Lens' ModifyVpnConnectionOptions (Maybe Text)

The IPv6 CIDR on the customer gateway (on-premises) side of the VPN connection.

Default: ::/0

modifyVpnConnectionOptions_dryRun :: Lens' ModifyVpnConnectionOptions (Maybe Bool)

Checks whether you have the required permissions for the action, without actually making the request, and provides an error response. If you have the required permissions, the error response is DryRunOperation. Otherwise, it is UnauthorizedOperation.

modifyVpnConnectionOptions_vpnConnectionId :: Lens' ModifyVpnConnectionOptions Text

The ID of the Site-to-Site VPN connection.

Destructuring the Response

data ModifyVpnConnectionOptionsResponse

See: newModifyVpnConnectionOptionsResponse smart constructor.

Constructors

ModifyVpnConnectionOptionsResponse'

Fields

vpnConnection :: Maybe VpnConnection

httpStatus :: Int

The response's http status code.

newModifyVpnConnectionOptionsResponse

Arguments

:: Int
$sel:httpStatus:ModifyVpnConnectionOptionsResponse'

-> ModifyVpnConnectionOptionsResponse

Create a value of ModifyVpnConnectionOptionsResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:vpnConnection:ModifyVpnConnectionOptionsResponse', modifyVpnConnectionOptionsResponse_vpnConnection - Undocumented member.

$sel:httpStatus:ModifyVpnConnectionOptionsResponse', modifyVpnConnectionOptionsResponse_httpStatus - The response's http status code.

Response Lenses

modifyVpnConnectionOptionsResponse_vpnConnection :: Lens' ModifyVpnConnectionOptionsResponse (Maybe VpnConnection)

Undocumented member.

modifyVpnConnectionOptionsResponse_httpStatus :: Lens' ModifyVpnConnectionOptionsResponse Int

The response's http status code.