Amazonka.EC2.PurchaseHostReservation

Description

Purchase a reservation with configurations that match those of your Dedicated Host. You must have active Dedicated Hosts in your account before you purchase a reservation. This action results in the specified reservation being purchased and charged to your account.

Creating a Request

data PurchaseHostReservation

See: newPurchaseHostReservation smart constructor.

Constructors

PurchaseHostReservation'

Fields

currencyCode :: Maybe CurrencyCodeValues

The currency in which the totalUpfrontPrice, LimitPrice, and totalHourlyPrice amounts are specified. At this time, the only supported currency is USD.

clientToken :: Maybe Text

Unique, case-sensitive identifier that you provide to ensure the idempotency of the request. For more information, see Ensuring Idempotency.

tagSpecifications :: Maybe [TagSpecification]

The tags to apply to the Dedicated Host Reservation during purchase.

limitPrice :: Maybe Text

The specified limit is checked against the total upfront cost of the reservation (calculated as the offering's upfront cost multiplied by the host count). If the total upfront cost is greater than the specified price limit, the request fails. This is used to ensure that the purchase does not exceed the expected upfront cost of the purchase. At this time, the only supported currency is USD. For example, to indicate a limit price of USD 100, specify 100.00.

hostIdSet :: [Text]

The IDs of the Dedicated Hosts with which the reservation will be associated.

offeringId :: Text

The ID of the offering.

newPurchaseHostReservation

Arguments

:: Text
$sel:offeringId:PurchaseHostReservation'

-> PurchaseHostReservation

Create a value of PurchaseHostReservation with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:currencyCode:PurchaseHostReservation', purchaseHostReservation_currencyCode - The currency in which the totalUpfrontPrice, LimitPrice, and totalHourlyPrice amounts are specified. At this time, the only supported currency is USD.

$sel:clientToken:PurchaseHostReservation', purchaseHostReservation_clientToken - Unique, case-sensitive identifier that you provide to ensure the idempotency of the request. For more information, see Ensuring Idempotency.

$sel:tagSpecifications:PurchaseHostReservation', purchaseHostReservation_tagSpecifications - The tags to apply to the Dedicated Host Reservation during purchase.

$sel:limitPrice:PurchaseHostReservation', purchaseHostReservation_limitPrice - The specified limit is checked against the total upfront cost of the reservation (calculated as the offering's upfront cost multiplied by the host count). If the total upfront cost is greater than the specified price limit, the request fails. This is used to ensure that the purchase does not exceed the expected upfront cost of the purchase. At this time, the only supported currency is USD. For example, to indicate a limit price of USD 100, specify 100.00.

$sel:hostIdSet:PurchaseHostReservation', purchaseHostReservation_hostIdSet - The IDs of the Dedicated Hosts with which the reservation will be associated.

$sel:offeringId:PurchaseHostReservation', purchaseHostReservation_offeringId - The ID of the offering.

Request Lenses

purchaseHostReservation_currencyCode :: Lens' PurchaseHostReservation (Maybe CurrencyCodeValues)

The currency in which the totalUpfrontPrice, LimitPrice, and totalHourlyPrice amounts are specified. At this time, the only supported currency is USD.

purchaseHostReservation_clientToken :: Lens' PurchaseHostReservation (Maybe Text)

Unique, case-sensitive identifier that you provide to ensure the idempotency of the request. For more information, see Ensuring Idempotency.

purchaseHostReservation_tagSpecifications :: Lens' PurchaseHostReservation (Maybe [TagSpecification])

The tags to apply to the Dedicated Host Reservation during purchase.

purchaseHostReservation_limitPrice :: Lens' PurchaseHostReservation (Maybe Text)

The specified limit is checked against the total upfront cost of the reservation (calculated as the offering's upfront cost multiplied by the host count). If the total upfront cost is greater than the specified price limit, the request fails. This is used to ensure that the purchase does not exceed the expected upfront cost of the purchase. At this time, the only supported currency is USD. For example, to indicate a limit price of USD 100, specify 100.00.

purchaseHostReservation_hostIdSet :: Lens' PurchaseHostReservation [Text]

The IDs of the Dedicated Hosts with which the reservation will be associated.

purchaseHostReservation_offeringId :: Lens' PurchaseHostReservation Text

The ID of the offering.

Destructuring the Response

data PurchaseHostReservationResponse

See: newPurchaseHostReservationResponse smart constructor.

Constructors

PurchaseHostReservationResponse'

Fields

currencyCode :: Maybe CurrencyCodeValues

The currency in which the totalUpfrontPrice and totalHourlyPrice amounts are specified. At this time, the only supported currency is USD.

clientToken :: Maybe Text

Unique, case-sensitive identifier that you provide to ensure the idempotency of the request. For more information, see Ensuring Idempotency.

totalHourlyPrice :: Maybe Text

The total hourly price of the reservation calculated per hour.

totalUpfrontPrice :: Maybe Text

The total amount charged to your account when you purchase the reservation.

purchase :: Maybe [Purchase]

Describes the details of the purchase.

httpStatus :: Int

The response's http status code.

newPurchaseHostReservationResponse

Arguments

:: Int
$sel:httpStatus:PurchaseHostReservationResponse'

-> PurchaseHostReservationResponse

Create a value of PurchaseHostReservationResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:currencyCode:PurchaseHostReservation', purchaseHostReservationResponse_currencyCode - The currency in which the totalUpfrontPrice and totalHourlyPrice amounts are specified. At this time, the only supported currency is USD.

$sel:clientToken:PurchaseHostReservation', purchaseHostReservationResponse_clientToken - Unique, case-sensitive identifier that you provide to ensure the idempotency of the request. For more information, see Ensuring Idempotency.

$sel:totalHourlyPrice:PurchaseHostReservationResponse', purchaseHostReservationResponse_totalHourlyPrice - The total hourly price of the reservation calculated per hour.

$sel:totalUpfrontPrice:PurchaseHostReservationResponse', purchaseHostReservationResponse_totalUpfrontPrice - The total amount charged to your account when you purchase the reservation.

$sel:purchase:PurchaseHostReservationResponse', purchaseHostReservationResponse_purchase - Describes the details of the purchase.

$sel:httpStatus:PurchaseHostReservationResponse', purchaseHostReservationResponse_httpStatus - The response's http status code.

Response Lenses

purchaseHostReservationResponse_currencyCode :: Lens' PurchaseHostReservationResponse (Maybe CurrencyCodeValues)

The currency in which the totalUpfrontPrice and totalHourlyPrice amounts are specified. At this time, the only supported currency is USD.

purchaseHostReservationResponse_clientToken :: Lens' PurchaseHostReservationResponse (Maybe Text)

Unique, case-sensitive identifier that you provide to ensure the idempotency of the request. For more information, see Ensuring Idempotency.

purchaseHostReservationResponse_totalHourlyPrice :: Lens' PurchaseHostReservationResponse (Maybe Text)

The total hourly price of the reservation calculated per hour.

purchaseHostReservationResponse_totalUpfrontPrice :: Lens' PurchaseHostReservationResponse (Maybe Text)

The total amount charged to your account when you purchase the reservation.

purchaseHostReservationResponse_purchase :: Lens' PurchaseHostReservationResponse (Maybe [Purchase])

Describes the details of the purchase.

purchaseHostReservationResponse_httpStatus :: Lens' PurchaseHostReservationResponse Int

The response's http status code.