Amazonka.EC2.ReleaseAddress

Description

Releases the specified Elastic IP address.

EC2-Classic, default VPC

Releasing an Elastic IP address automatically disassociates it from any instance that it's associated with. To disassociate an Elastic IP address without releasing it, use DisassociateAddress.

Nondefault VPC

You must use DisassociateAddress to disassociate the Elastic IP address before you can release it. Otherwise, Amazon EC2 returns an error (InvalidIPAddress.InUse).

After releasing an Elastic IP address, it is released to the IP address pool. Be sure to update your DNS records and any servers or devices that communicate with the address. If you attempt to release an Elastic IP address that you already released, you'll get an AuthFailure error if the address is already allocated to another Amazon Web Services account.

EC2-VPC

After you release an Elastic IP address for use in a VPC, you might be able to recover it. For more information, see AllocateAddress.

Creating a Request

data ReleaseAddress

See: newReleaseAddress smart constructor.

Constructors

ReleaseAddress'

Fields

allocationId :: Maybe Text

EC2-VPC

The allocation ID. Required for EC2-VPC.

networkBorderGroup :: Maybe Text

The set of Availability Zones, Local Zones, or Wavelength Zones from which Amazon Web Services advertises IP addresses.

If you provide an incorrect network border group, you receive an InvalidAddress.NotFound error.

You cannot use a network border group with EC2 Classic. If you attempt this operation on EC2 classic, you receive an InvalidParameterCombination error.

publicIp :: Maybe Text

EC2-Classic

The Elastic IP address. Required for EC2-Classic.

dryRun :: Maybe Bool

Checks whether you have the required permissions for the action, without actually making the request, and provides an error response. If you have the required permissions, the error response is DryRunOperation. Otherwise, it is UnauthorizedOperation.

newReleaseAddress :: ReleaseAddress

Create a value of ReleaseAddress with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:allocationId:ReleaseAddress', releaseAddress_allocationId - [EC2-VPC] The allocation ID. Required for EC2-VPC.

$sel:networkBorderGroup:ReleaseAddress', releaseAddress_networkBorderGroup - The set of Availability Zones, Local Zones, or Wavelength Zones from which Amazon Web Services advertises IP addresses.

If you provide an incorrect network border group, you receive an InvalidAddress.NotFound error.

You cannot use a network border group with EC2 Classic. If you attempt this operation on EC2 classic, you receive an InvalidParameterCombination error.

$sel:publicIp:ReleaseAddress', releaseAddress_publicIp - [EC2-Classic] The Elastic IP address. Required for EC2-Classic.

$sel:dryRun:ReleaseAddress', releaseAddress_dryRun - Checks whether you have the required permissions for the action, without actually making the request, and provides an error response. If you have the required permissions, the error response is DryRunOperation. Otherwise, it is UnauthorizedOperation.

Request Lenses

releaseAddress_allocationId :: Lens' ReleaseAddress (Maybe Text)

EC2-VPC

The allocation ID. Required for EC2-VPC.

releaseAddress_networkBorderGroup :: Lens' ReleaseAddress (Maybe Text)

The set of Availability Zones, Local Zones, or Wavelength Zones from which Amazon Web Services advertises IP addresses.

If you provide an incorrect network border group, you receive an InvalidAddress.NotFound error.

You cannot use a network border group with EC2 Classic. If you attempt this operation on EC2 classic, you receive an InvalidParameterCombination error.

releaseAddress_publicIp :: Lens' ReleaseAddress (Maybe Text)

EC2-Classic

The Elastic IP address. Required for EC2-Classic.

releaseAddress_dryRun :: Lens' ReleaseAddress (Maybe Bool)

Checks whether you have the required permissions for the action, without actually making the request, and provides an error response. If you have the required permissions, the error response is DryRunOperation. Otherwise, it is UnauthorizedOperation.

Destructuring the Response

data ReleaseAddressResponse

See: newReleaseAddressResponse smart constructor.

Constructors

ReleaseAddressResponse'

newReleaseAddressResponse :: ReleaseAddressResponse

Create a value of ReleaseAddressResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.