Amazonka.EC2.ReportInstanceStatus

Description

Submits feedback about the status of an instance. The instance must be in the running state. If your experience with the instance differs from the instance status returned by DescribeInstanceStatus, use ReportInstanceStatus to report your experience with the instance. Amazon EC2 collects this information to improve the accuracy of status checks.

Use of this action does not change the value returned by DescribeInstanceStatus.

Creating a Request

data ReportInstanceStatus

See: newReportInstanceStatus smart constructor.

Constructors

ReportInstanceStatus'

Fields

startTime :: Maybe ISO8601

The time at which the reported instance health state began.

endTime :: Maybe ISO8601

The time at which the reported instance health state ended.

description :: Maybe Text

Descriptive text about the health state of your instance.

dryRun :: Maybe Bool

Checks whether you have the required permissions for the action, without actually making the request, and provides an error response. If you have the required permissions, the error response is DryRunOperation. Otherwise, it is UnauthorizedOperation.

instances :: [Text]

The instances.

reasonCodes :: [ReportInstanceReasonCodes]

The reason codes that describe the health state of your instance.

instance-stuck-in-state: My instance is stuck in a state.

unresponsive: My instance is unresponsive.

not-accepting-credentials: My instance is not accepting my credentials.

password-not-available: A password is not available for my instance.

performance-network: My instance is experiencing performance problems that I believe are network related.

performance-instance-store: My instance is experiencing performance problems that I believe are related to the instance stores.

performance-ebs-volume: My instance is experiencing performance problems that I believe are related to an EBS volume.

performance-other: My instance is experiencing performance problems.

other: [explain using the description parameter]

status :: ReportStatusType

The status of all instances listed.

newReportInstanceStatus

Arguments

:: ReportStatusType
$sel:status:ReportInstanceStatus'

-> ReportInstanceStatus

Create a value of ReportInstanceStatus with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:startTime:ReportInstanceStatus', reportInstanceStatus_startTime - The time at which the reported instance health state began.

$sel:endTime:ReportInstanceStatus', reportInstanceStatus_endTime - The time at which the reported instance health state ended.

$sel:description:ReportInstanceStatus', reportInstanceStatus_description - Descriptive text about the health state of your instance.

$sel:dryRun:ReportInstanceStatus', reportInstanceStatus_dryRun - Checks whether you have the required permissions for the action, without actually making the request, and provides an error response. If you have the required permissions, the error response is DryRunOperation. Otherwise, it is UnauthorizedOperation.

$sel:instances:ReportInstanceStatus', reportInstanceStatus_instances - The instances.

$sel:reasonCodes:ReportInstanceStatus', reportInstanceStatus_reasonCodes - The reason codes that describe the health state of your instance.

instance-stuck-in-state: My instance is stuck in a state.

unresponsive: My instance is unresponsive.

not-accepting-credentials: My instance is not accepting my credentials.

password-not-available: A password is not available for my instance.

performance-network: My instance is experiencing performance problems that I believe are network related.

performance-instance-store: My instance is experiencing performance problems that I believe are related to the instance stores.

performance-ebs-volume: My instance is experiencing performance problems that I believe are related to an EBS volume.

performance-other: My instance is experiencing performance problems.

other: [explain using the description parameter]

$sel:status:ReportInstanceStatus', reportInstanceStatus_status - The status of all instances listed.

Request Lenses

reportInstanceStatus_startTime :: Lens' ReportInstanceStatus (Maybe UTCTime)

The time at which the reported instance health state began.

reportInstanceStatus_endTime :: Lens' ReportInstanceStatus (Maybe UTCTime)

The time at which the reported instance health state ended.

reportInstanceStatus_description :: Lens' ReportInstanceStatus (Maybe Text)

Descriptive text about the health state of your instance.

reportInstanceStatus_dryRun :: Lens' ReportInstanceStatus (Maybe Bool)

Checks whether you have the required permissions for the action, without actually making the request, and provides an error response. If you have the required permissions, the error response is DryRunOperation. Otherwise, it is UnauthorizedOperation.

reportInstanceStatus_instances :: Lens' ReportInstanceStatus [Text]

The instances.

reportInstanceStatus_reasonCodes :: Lens' ReportInstanceStatus [ReportInstanceReasonCodes]

The reason codes that describe the health state of your instance.

instance-stuck-in-state: My instance is stuck in a state.

unresponsive: My instance is unresponsive.

not-accepting-credentials: My instance is not accepting my credentials.

password-not-available: A password is not available for my instance.

performance-network: My instance is experiencing performance problems that I believe are network related.

performance-instance-store: My instance is experiencing performance problems that I believe are related to the instance stores.

performance-ebs-volume: My instance is experiencing performance problems that I believe are related to an EBS volume.

performance-other: My instance is experiencing performance problems.

other: [explain using the description parameter]

reportInstanceStatus_status :: Lens' ReportInstanceStatus ReportStatusType

The status of all instances listed.

Destructuring the Response

data ReportInstanceStatusResponse

See: newReportInstanceStatusResponse smart constructor.

Constructors

ReportInstanceStatusResponse'

newReportInstanceStatusResponse :: ReportInstanceStatusResponse

Create a value of ReportInstanceStatusResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.