Amazonka.EC2.ResetAddressAttribute

Description

Resets the attribute of the specified IP address. For requirements, see Using reverse DNS for email applications.

Creating a Request

data ResetAddressAttribute

See: newResetAddressAttribute smart constructor.

Constructors

ResetAddressAttribute'

Fields

dryRun :: Maybe Bool

Checks whether you have the required permissions for the action, without actually making the request, and provides an error response. If you have the required permissions, the error response is DryRunOperation. Otherwise, it is UnauthorizedOperation.

allocationId :: Text

EC2-VPC

The allocation ID.

attribute :: AddressAttributeName

The attribute of the IP address.

newResetAddressAttribute

Arguments

:: Text
$sel:allocationId:ResetAddressAttribute'

-> AddressAttributeName
$sel:attribute:ResetAddressAttribute'

-> ResetAddressAttribute

Create a value of ResetAddressAttribute with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:dryRun:ResetAddressAttribute', resetAddressAttribute_dryRun - Checks whether you have the required permissions for the action, without actually making the request, and provides an error response. If you have the required permissions, the error response is DryRunOperation. Otherwise, it is UnauthorizedOperation.

$sel:allocationId:ResetAddressAttribute', resetAddressAttribute_allocationId - [EC2-VPC] The allocation ID.

$sel:attribute:ResetAddressAttribute', resetAddressAttribute_attribute - The attribute of the IP address.

Request Lenses

resetAddressAttribute_dryRun :: Lens' ResetAddressAttribute (Maybe Bool)

Checks whether you have the required permissions for the action, without actually making the request, and provides an error response. If you have the required permissions, the error response is DryRunOperation. Otherwise, it is UnauthorizedOperation.

resetAddressAttribute_allocationId :: Lens' ResetAddressAttribute Text

EC2-VPC

The allocation ID.

resetAddressAttribute_attribute :: Lens' ResetAddressAttribute AddressAttributeName

The attribute of the IP address.

Destructuring the Response

data ResetAddressAttributeResponse

See: newResetAddressAttributeResponse smart constructor.

Constructors

ResetAddressAttributeResponse'

Fields

address :: Maybe AddressAttribute

Information about the IP address.

httpStatus :: Int

The response's http status code.

newResetAddressAttributeResponse

Arguments

:: Int
$sel:httpStatus:ResetAddressAttributeResponse'

-> ResetAddressAttributeResponse

Create a value of ResetAddressAttributeResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:address:ResetAddressAttributeResponse', resetAddressAttributeResponse_address - Information about the IP address.

$sel:httpStatus:ResetAddressAttributeResponse', resetAddressAttributeResponse_httpStatus - The response's http status code.

Response Lenses

resetAddressAttributeResponse_address :: Lens' ResetAddressAttributeResponse (Maybe AddressAttribute)

Information about the IP address.

resetAddressAttributeResponse_httpStatus :: Lens' ResetAddressAttributeResponse Int

The response's http status code.