Amazonka.EC2.RunScheduledInstances

Description

Launches the specified Scheduled Instances.

Before you can launch a Scheduled Instance, you must purchase it and obtain an identifier using PurchaseScheduledInstances.

You must launch a Scheduled Instance during its scheduled time period. You can't stop or reboot a Scheduled Instance, but you can terminate it as needed. If you terminate a Scheduled Instance before the current scheduled time period ends, you can launch it again after a few minutes. For more information, see Scheduled Instances in the Amazon EC2 User Guide.

Creating a Request

data RunScheduledInstances

Contains the parameters for RunScheduledInstances.

See: newRunScheduledInstances smart constructor.

Constructors

RunScheduledInstances'

Fields

clientToken :: Maybe Text

Unique, case-sensitive identifier that ensures the idempotency of the request. For more information, see Ensuring Idempotency.

instanceCount :: Maybe Int

The number of instances.

Default: 1

dryRun :: Maybe Bool

Checks whether you have the required permissions for the action, without actually making the request, and provides an error response. If you have the required permissions, the error response is DryRunOperation. Otherwise, it is UnauthorizedOperation.

launchSpecification :: ScheduledInstancesLaunchSpecification

The launch specification. You must match the instance type, Availability Zone, network, and platform of the schedule that you purchased.

scheduledInstanceId :: Text

The Scheduled Instance ID.

newRunScheduledInstances

Arguments

:: ScheduledInstancesLaunchSpecification
$sel:launchSpecification:RunScheduledInstances'

-> Text
$sel:scheduledInstanceId:RunScheduledInstances'

-> RunScheduledInstances

Create a value of RunScheduledInstances with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:clientToken:RunScheduledInstances', runScheduledInstances_clientToken - Unique, case-sensitive identifier that ensures the idempotency of the request. For more information, see Ensuring Idempotency.

$sel:instanceCount:RunScheduledInstances', runScheduledInstances_instanceCount - The number of instances.

Default: 1

$sel:dryRun:RunScheduledInstances', runScheduledInstances_dryRun - Checks whether you have the required permissions for the action, without actually making the request, and provides an error response. If you have the required permissions, the error response is DryRunOperation. Otherwise, it is UnauthorizedOperation.

$sel:launchSpecification:RunScheduledInstances', runScheduledInstances_launchSpecification - The launch specification. You must match the instance type, Availability Zone, network, and platform of the schedule that you purchased.

$sel:scheduledInstanceId:RunScheduledInstances', runScheduledInstances_scheduledInstanceId - The Scheduled Instance ID.

Request Lenses

runScheduledInstances_clientToken :: Lens' RunScheduledInstances (Maybe Text)

Unique, case-sensitive identifier that ensures the idempotency of the request. For more information, see Ensuring Idempotency.

runScheduledInstances_instanceCount :: Lens' RunScheduledInstances (Maybe Int)

The number of instances.

Default: 1

runScheduledInstances_dryRun :: Lens' RunScheduledInstances (Maybe Bool)

Checks whether you have the required permissions for the action, without actually making the request, and provides an error response. If you have the required permissions, the error response is DryRunOperation. Otherwise, it is UnauthorizedOperation.

runScheduledInstances_launchSpecification :: Lens' RunScheduledInstances ScheduledInstancesLaunchSpecification

The launch specification. You must match the instance type, Availability Zone, network, and platform of the schedule that you purchased.

runScheduledInstances_scheduledInstanceId :: Lens' RunScheduledInstances Text

The Scheduled Instance ID.

Destructuring the Response

data RunScheduledInstancesResponse

Contains the output of RunScheduledInstances.

See: newRunScheduledInstancesResponse smart constructor.

Constructors

RunScheduledInstancesResponse'

Fields

instanceIdSet :: Maybe [Text]

The IDs of the newly launched instances.

httpStatus :: Int

The response's http status code.

newRunScheduledInstancesResponse

Arguments

:: Int
$sel:httpStatus:RunScheduledInstancesResponse'

-> RunScheduledInstancesResponse

Create a value of RunScheduledInstancesResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:instanceIdSet:RunScheduledInstancesResponse', runScheduledInstancesResponse_instanceIdSet - The IDs of the newly launched instances.

$sel:httpStatus:RunScheduledInstancesResponse', runScheduledInstancesResponse_httpStatus - The response's http status code.

Response Lenses

runScheduledInstancesResponse_instanceIdSet :: Lens' RunScheduledInstancesResponse (Maybe [Text])

The IDs of the newly launched instances.

runScheduledInstancesResponse_httpStatus :: Lens' RunScheduledInstancesResponse Int

The response's http status code.