Amazonka.EC2.StartInstances

Description

Starts an Amazon EBS-backed instance that you've previously stopped.

Instances that use Amazon EBS volumes as their root devices can be quickly stopped and started. When an instance is stopped, the compute resources are released and you are not billed for instance usage. However, your root partition Amazon EBS volume remains and continues to persist your data, and you are charged for Amazon EBS volume usage. You can restart your instance at any time. Every time you start your instance, Amazon EC2 charges a one-minute minimum for instance usage, and thereafter charges per second for instance usage.

Before stopping an instance, make sure it is in a state from which it can be restarted. Stopping an instance does not preserve data stored in RAM.

Performing this operation on an instance that uses an instance store as its root device returns an error.

If you attempt to start a T3 instance with host tenancy and the unlimted CPU credit option, the request fails. The unlimited CPU credit option is not supported on Dedicated Hosts. Before you start the instance, either change its CPU credit option to standard, or change its tenancy to default or dedicated.

For more information, see Stopping instances in the Amazon EC2 User Guide.

Creating a Request

data StartInstances

See: newStartInstances smart constructor.

Constructors

StartInstances'

Fields

additionalInfo :: Maybe Text

Reserved.

dryRun :: Maybe Bool

Checks whether you have the required permissions for the action, without actually making the request, and provides an error response. If you have the required permissions, the error response is DryRunOperation. Otherwise, it is UnauthorizedOperation.

instanceIds :: [Text]

The IDs of the instances.

newStartInstances :: StartInstances

Create a value of StartInstances with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:additionalInfo:StartInstances', startInstances_additionalInfo - Reserved.

$sel:dryRun:StartInstances', startInstances_dryRun - Checks whether you have the required permissions for the action, without actually making the request, and provides an error response. If you have the required permissions, the error response is DryRunOperation. Otherwise, it is UnauthorizedOperation.

$sel:instanceIds:StartInstances', startInstances_instanceIds - The IDs of the instances.

Request Lenses

startInstances_additionalInfo :: Lens' StartInstances (Maybe Text)

Reserved.

startInstances_dryRun :: Lens' StartInstances (Maybe Bool)

Checks whether you have the required permissions for the action, without actually making the request, and provides an error response. If you have the required permissions, the error response is DryRunOperation. Otherwise, it is UnauthorizedOperation.

startInstances_instanceIds :: Lens' StartInstances [Text]

The IDs of the instances.

Destructuring the Response

data StartInstancesResponse

See: newStartInstancesResponse smart constructor.

Constructors

StartInstancesResponse'

Fields

startingInstances :: Maybe [InstanceStateChange]

Information about the started instances.

httpStatus :: Int

The response's http status code.

newStartInstancesResponse

Arguments

:: Int
$sel:httpStatus:StartInstancesResponse'

-> StartInstancesResponse

Create a value of StartInstancesResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:startingInstances:StartInstancesResponse', startInstancesResponse_startingInstances - Information about the started instances.

$sel:httpStatus:StartInstancesResponse', startInstancesResponse_httpStatus - The response's http status code.

Response Lenses

startInstancesResponse_startingInstances :: Lens' StartInstancesResponse (Maybe [InstanceStateChange])

Information about the started instances.

startInstancesResponse_httpStatus :: Lens' StartInstancesResponse Int

The response's http status code.