Amazonka.EC2.TerminateInstances

Description

Shuts down the specified instances. This operation is idempotent; if you terminate an instance more than once, each call succeeds.

If you specify multiple instances and the request fails (for example, because of a single incorrect instance ID), none of the instances are terminated.

If you terminate multiple instances across multiple Availability Zones, and one or more of the specified instances are enabled for termination protection, the request fails with the following results:

The specified instances that are in the same Availability Zone as the protected instance are not terminated.

The specified instances that are in different Availability Zones, where no other specified instances are protected, are successfully terminated.

For example, say you have the following instances:

Instance A: us-east-1a; Not protected

Instance B: us-east-1a; Not protected

Instance C: us-east-1b; Protected

Instance D: us-east-1b; not protected

If you attempt to terminate all of these instances in the same request, the request reports failure with the following results:

Instance A and Instance B are successfully terminated because none of the specified instances in us-east-1a are enabled for termination protection.

Instance C and Instance D fail to terminate because at least one of the specified instances in us-east-1b (Instance C) is enabled for termination protection.

Terminated instances remain visible after termination (for approximately one hour).

By default, Amazon EC2 deletes all EBS volumes that were attached when the instance launched. Volumes attached after instance launch continue running.

You can stop, start, and terminate EBS-backed instances. You can only terminate instance store-backed instances. What happens to an instance differs if you stop it or terminate it. For example, when you stop an instance, the root device and any other devices attached to the instance persist. When you terminate an instance, any attached EBS volumes with the DeleteOnTermination block device mapping parameter set to true are automatically deleted. For more information about the differences between stopping and terminating instances, see Instance lifecycle in the Amazon EC2 User Guide.

For more information about troubleshooting, see Troubleshooting terminating your instance in the Amazon EC2 User Guide.

Creating a Request

data TerminateInstances

See: newTerminateInstances smart constructor.

Constructors

TerminateInstances'

Fields

dryRun :: Maybe Bool

Checks whether you have the required permissions for the action, without actually making the request, and provides an error response. If you have the required permissions, the error response is DryRunOperation. Otherwise, it is UnauthorizedOperation.

instanceIds :: [Text]

The IDs of the instances.

Constraints: Up to 1000 instance IDs. We recommend breaking up this request into smaller batches.

newTerminateInstances :: TerminateInstances

Create a value of TerminateInstances with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:dryRun:TerminateInstances', terminateInstances_dryRun - Checks whether you have the required permissions for the action, without actually making the request, and provides an error response. If you have the required permissions, the error response is DryRunOperation. Otherwise, it is UnauthorizedOperation.

$sel:instanceIds:TerminateInstances', terminateInstances_instanceIds - The IDs of the instances.

Constraints: Up to 1000 instance IDs. We recommend breaking up this request into smaller batches.

Request Lenses

terminateInstances_dryRun :: Lens' TerminateInstances (Maybe Bool)

Checks whether you have the required permissions for the action, without actually making the request, and provides an error response. If you have the required permissions, the error response is DryRunOperation. Otherwise, it is UnauthorizedOperation.

terminateInstances_instanceIds :: Lens' TerminateInstances [Text]

The IDs of the instances.

Constraints: Up to 1000 instance IDs. We recommend breaking up this request into smaller batches.

Destructuring the Response

data TerminateInstancesResponse

See: newTerminateInstancesResponse smart constructor.

Constructors

TerminateInstancesResponse'

Fields

terminatingInstances :: Maybe [InstanceStateChange]

Information about the terminated instances.

httpStatus :: Int

The response's http status code.

newTerminateInstancesResponse

Arguments

:: Int
$sel:httpStatus:TerminateInstancesResponse'

-> TerminateInstancesResponse

Create a value of TerminateInstancesResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:terminatingInstances:TerminateInstancesResponse', terminateInstancesResponse_terminatingInstances - Information about the terminated instances.

$sel:httpStatus:TerminateInstancesResponse', terminateInstancesResponse_httpStatus - The response's http status code.

Response Lenses

terminateInstancesResponse_terminatingInstances :: Lens' TerminateInstancesResponse (Maybe [InstanceStateChange])

Information about the terminated instances.

terminateInstancesResponse_httpStatus :: Lens' TerminateInstancesResponse Int

The response's http status code.