Amazonka.EC2.Types.ActiveInstance

Description

Documentation

data ActiveInstance

Describes a running instance in a Spot Fleet.

See: newActiveInstance smart constructor.

Constructors

ActiveInstance'

Fields

instanceId :: Maybe Text

The ID of the instance.

instanceHealth :: Maybe InstanceHealthStatus

The health status of the instance. If the status of either the instance status check or the system status check is impaired, the health status of the instance is unhealthy. Otherwise, the health status is healthy.

instanceType :: Maybe Text

The instance type.

spotInstanceRequestId :: Maybe Text

The ID of the Spot Instance request.

newActiveInstance :: ActiveInstance

Create a value of ActiveInstance with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:instanceId:ActiveInstance', activeInstance_instanceId - The ID of the instance.

$sel:instanceHealth:ActiveInstance', activeInstance_instanceHealth - The health status of the instance. If the status of either the instance status check or the system status check is impaired, the health status of the instance is unhealthy. Otherwise, the health status is healthy.

$sel:instanceType:ActiveInstance', activeInstance_instanceType - The instance type.

$sel:spotInstanceRequestId:ActiveInstance', activeInstance_spotInstanceRequestId - The ID of the Spot Instance request.

activeInstance_instanceId :: Lens' ActiveInstance (Maybe Text)

The ID of the instance.

activeInstance_instanceHealth :: Lens' ActiveInstance (Maybe InstanceHealthStatus)

The health status of the instance. If the status of either the instance status check or the system status check is impaired, the health status of the instance is unhealthy. Otherwise, the health status is healthy.

activeInstance_instanceType :: Lens' ActiveInstance (Maybe Text)

The instance type.

activeInstance_spotInstanceRequestId :: Lens' ActiveInstance (Maybe Text)

The ID of the Spot Instance request.