Amazonka.EC2.Types.AnalysisPacketHeader

Description

Documentation

data AnalysisPacketHeader

Describes a header. Reflects any changes made by a component as traffic passes through. The fields of an inbound header are null except for the first component of a path.

See: newAnalysisPacketHeader smart constructor.

Constructors

AnalysisPacketHeader'

Fields

destinationAddresses :: Maybe [Text]

The destination addresses.

sourceAddresses :: Maybe [Text]

The source addresses.

protocol :: Maybe Text

The protocol.

destinationPortRanges :: Maybe [PortRange]

The destination port ranges.

sourcePortRanges :: Maybe [PortRange]

The source port ranges.

newAnalysisPacketHeader :: AnalysisPacketHeader

Create a value of AnalysisPacketHeader with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:destinationAddresses:AnalysisPacketHeader', analysisPacketHeader_destinationAddresses - The destination addresses.

$sel:sourceAddresses:AnalysisPacketHeader', analysisPacketHeader_sourceAddresses - The source addresses.

$sel:protocol:AnalysisPacketHeader', analysisPacketHeader_protocol - The protocol.

$sel:destinationPortRanges:AnalysisPacketHeader', analysisPacketHeader_destinationPortRanges - The destination port ranges.

$sel:sourcePortRanges:AnalysisPacketHeader', analysisPacketHeader_sourcePortRanges - The source port ranges.

analysisPacketHeader_destinationAddresses :: Lens' AnalysisPacketHeader (Maybe [Text])

The destination addresses.

analysisPacketHeader_sourceAddresses :: Lens' AnalysisPacketHeader (Maybe [Text])

The source addresses.

analysisPacketHeader_protocol :: Lens' AnalysisPacketHeader (Maybe Text)

The protocol.

analysisPacketHeader_destinationPortRanges :: Lens' AnalysisPacketHeader (Maybe [PortRange])

The destination port ranges.

analysisPacketHeader_sourcePortRanges :: Lens' AnalysisPacketHeader (Maybe [PortRange])

The source port ranges.