Amazonka.EC2.Types.AuthorizationRule

Description

Documentation

data AuthorizationRule

Information about an authorization rule.

See: newAuthorizationRule smart constructor.

Constructors

AuthorizationRule'

Fields

status :: Maybe ClientVpnAuthorizationRuleStatus

The current state of the authorization rule.

accessAll :: Maybe Bool

Indicates whether the authorization rule grants access to all clients.

clientVpnEndpointId :: Maybe Text

The ID of the Client VPN endpoint with which the authorization rule is associated.

groupId :: Maybe Text

The ID of the Active Directory group to which the authorization rule grants access.

destinationCidr :: Maybe Text

The IPv4 address range, in CIDR notation, of the network to which the authorization rule applies.

description :: Maybe Text

A brief description of the authorization rule.

newAuthorizationRule :: AuthorizationRule

Create a value of AuthorizationRule with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:status:AuthorizationRule', authorizationRule_status - The current state of the authorization rule.

$sel:accessAll:AuthorizationRule', authorizationRule_accessAll - Indicates whether the authorization rule grants access to all clients.

$sel:clientVpnEndpointId:AuthorizationRule', authorizationRule_clientVpnEndpointId - The ID of the Client VPN endpoint with which the authorization rule is associated.

$sel:groupId:AuthorizationRule', authorizationRule_groupId - The ID of the Active Directory group to which the authorization rule grants access.

$sel:destinationCidr:AuthorizationRule', authorizationRule_destinationCidr - The IPv4 address range, in CIDR notation, of the network to which the authorization rule applies.

$sel:description:AuthorizationRule', authorizationRule_description - A brief description of the authorization rule.

authorizationRule_status :: Lens' AuthorizationRule (Maybe ClientVpnAuthorizationRuleStatus)

The current state of the authorization rule.

authorizationRule_accessAll :: Lens' AuthorizationRule (Maybe Bool)

Indicates whether the authorization rule grants access to all clients.

authorizationRule_clientVpnEndpointId :: Lens' AuthorizationRule (Maybe Text)

The ID of the Client VPN endpoint with which the authorization rule is associated.

authorizationRule_groupId :: Lens' AuthorizationRule (Maybe Text)

The ID of the Active Directory group to which the authorization rule grants access.

authorizationRule_destinationCidr :: Lens' AuthorizationRule (Maybe Text)

The IPv4 address range, in CIDR notation, of the network to which the authorization rule applies.

authorizationRule_description :: Lens' AuthorizationRule (Maybe Text)

A brief description of the authorization rule.