Amazonka.EC2.Types.CapacityReservationFleet

Description

Documentation

data CapacityReservationFleet

Information about a Capacity Reservation Fleet.

See: newCapacityReservationFleet smart constructor.

Constructors

CapacityReservationFleet'

Fields

capacityReservationFleetId :: Maybe Text

The ID of the Capacity Reservation Fleet.

state :: Maybe CapacityReservationFleetState

The state of the Capacity Reservation Fleet. Possible states include:

submitted - The Capacity Reservation Fleet request has been submitted and Amazon Elastic Compute Cloud is preparing to create the Capacity Reservations.

modifying - The Capacity Reservation Fleet is being modified. The Fleet remains in this state until the modification is complete.

active - The Capacity Reservation Fleet has fulfilled its total target capacity and it is attempting to maintain this capacity. The Fleet remains in this state until it is modified or deleted.

partially_fulfilled - The Capacity Reservation Fleet has partially fulfilled its total target capacity. There is insufficient Amazon EC2 to fulfill the total target capacity. The Fleet is attempting to asynchronously fulfill its total target capacity.

expiring - The Capacity Reservation Fleet has reach its end date and it is in the process of expiring. One or more of its Capacity reservations might still be active.

expired - The Capacity Reservation Fleet has reach its end date. The Fleet and its Capacity Reservations are expired. The Fleet can't create new Capacity Reservations.

cancelling - The Capacity Reservation Fleet is in the process of being cancelled. One or more of its Capacity reservations might still be active.

cancelled - The Capacity Reservation Fleet has been manually cancelled. The Fleet and its Capacity Reservations are cancelled and the Fleet can't create new Capacity Reservations.

failed - The Capacity Reservation Fleet failed to reserve capacity for the specified instance types.

totalFulfilledCapacity :: Maybe Double

The capacity units that have been fulfilled.

endDate :: Maybe ISO8601

The date and time at which the Capacity Reservation Fleet expires.

instanceMatchCriteria :: Maybe FleetInstanceMatchCriteria

Indicates the type of instance launches that the Capacity Reservation Fleet accepts. All Capacity Reservations in the Fleet inherit this instance matching criteria.

Currently, Capacity Reservation Fleets support open instance matching criteria only. This means that instances that have matching attributes (instance type, platform, and Availability Zone) run in the Capacity Reservations automatically. Instances do not need to explicitly target a Capacity Reservation Fleet to use its reserved capacity.

totalTargetCapacity :: Maybe Int

The total number of capacity units for which the Capacity Reservation Fleet reserves capacity. For more information, see Total target capacity in the Amazon EC2 User Guide.

capacityReservationFleetArn :: Maybe Text

The ARN of the Capacity Reservation Fleet.

instanceTypeSpecifications :: Maybe [FleetCapacityReservation]

Information about the instance types for which to reserve the capacity.

tenancy :: Maybe FleetCapacityReservationTenancy

The tenancy of the Capacity Reservation Fleet. Tenancies include:

default - The Capacity Reservation Fleet is created on hardware that is shared with other Amazon Web Services accounts.

dedicated - The Capacity Reservation Fleet is created on single-tenant hardware that is dedicated to a single Amazon Web Services account.

allocationStrategy :: Maybe Text

The strategy used by the Capacity Reservation Fleet to determine which of the specified instance types to use. For more information, see For more information, see Allocation strategy in the Amazon EC2 User Guide.

createTime :: Maybe ISO8601

The date and time at which the Capacity Reservation Fleet was created.

tags :: Maybe [Tag]

The tags assigned to the Capacity Reservation Fleet.

newCapacityReservationFleet :: CapacityReservationFleet

Create a value of CapacityReservationFleet with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:capacityReservationFleetId:CapacityReservationFleet', capacityReservationFleet_capacityReservationFleetId - The ID of the Capacity Reservation Fleet.

$sel:state:CapacityReservationFleet', capacityReservationFleet_state - The state of the Capacity Reservation Fleet. Possible states include:

submitted - The Capacity Reservation Fleet request has been submitted and Amazon Elastic Compute Cloud is preparing to create the Capacity Reservations.

modifying - The Capacity Reservation Fleet is being modified. The Fleet remains in this state until the modification is complete.

active - The Capacity Reservation Fleet has fulfilled its total target capacity and it is attempting to maintain this capacity. The Fleet remains in this state until it is modified or deleted.

partially_fulfilled - The Capacity Reservation Fleet has partially fulfilled its total target capacity. There is insufficient Amazon EC2 to fulfill the total target capacity. The Fleet is attempting to asynchronously fulfill its total target capacity.

expiring - The Capacity Reservation Fleet has reach its end date and it is in the process of expiring. One or more of its Capacity reservations might still be active.

expired - The Capacity Reservation Fleet has reach its end date. The Fleet and its Capacity Reservations are expired. The Fleet can't create new Capacity Reservations.

cancelling - The Capacity Reservation Fleet is in the process of being cancelled. One or more of its Capacity reservations might still be active.

cancelled - The Capacity Reservation Fleet has been manually cancelled. The Fleet and its Capacity Reservations are cancelled and the Fleet can't create new Capacity Reservations.

failed - The Capacity Reservation Fleet failed to reserve capacity for the specified instance types.

$sel:totalFulfilledCapacity:CapacityReservationFleet', capacityReservationFleet_totalFulfilledCapacity - The capacity units that have been fulfilled.

$sel:endDate:CapacityReservationFleet', capacityReservationFleet_endDate - The date and time at which the Capacity Reservation Fleet expires.

$sel:instanceMatchCriteria:CapacityReservationFleet', capacityReservationFleet_instanceMatchCriteria - Indicates the type of instance launches that the Capacity Reservation Fleet accepts. All Capacity Reservations in the Fleet inherit this instance matching criteria.

Currently, Capacity Reservation Fleets support open instance matching criteria only. This means that instances that have matching attributes (instance type, platform, and Availability Zone) run in the Capacity Reservations automatically. Instances do not need to explicitly target a Capacity Reservation Fleet to use its reserved capacity.

$sel:totalTargetCapacity:CapacityReservationFleet', capacityReservationFleet_totalTargetCapacity - The total number of capacity units for which the Capacity Reservation Fleet reserves capacity. For more information, see Total target capacity in the Amazon EC2 User Guide.

$sel:capacityReservationFleetArn:CapacityReservationFleet', capacityReservationFleet_capacityReservationFleetArn - The ARN of the Capacity Reservation Fleet.

$sel:instanceTypeSpecifications:CapacityReservationFleet', capacityReservationFleet_instanceTypeSpecifications - Information about the instance types for which to reserve the capacity.

$sel:tenancy:CapacityReservationFleet', capacityReservationFleet_tenancy - The tenancy of the Capacity Reservation Fleet. Tenancies include:

default - The Capacity Reservation Fleet is created on hardware that is shared with other Amazon Web Services accounts.

dedicated - The Capacity Reservation Fleet is created on single-tenant hardware that is dedicated to a single Amazon Web Services account.

$sel:allocationStrategy:CapacityReservationFleet', capacityReservationFleet_allocationStrategy - The strategy used by the Capacity Reservation Fleet to determine which of the specified instance types to use. For more information, see For more information, see Allocation strategy in the Amazon EC2 User Guide.

$sel:createTime:CapacityReservationFleet', capacityReservationFleet_createTime - The date and time at which the Capacity Reservation Fleet was created.

$sel:tags:CapacityReservationFleet', capacityReservationFleet_tags - The tags assigned to the Capacity Reservation Fleet.

capacityReservationFleet_capacityReservationFleetId :: Lens' CapacityReservationFleet (Maybe Text)

The ID of the Capacity Reservation Fleet.

capacityReservationFleet_state :: Lens' CapacityReservationFleet (Maybe CapacityReservationFleetState)

The state of the Capacity Reservation Fleet. Possible states include:

submitted - The Capacity Reservation Fleet request has been submitted and Amazon Elastic Compute Cloud is preparing to create the Capacity Reservations.

modifying - The Capacity Reservation Fleet is being modified. The Fleet remains in this state until the modification is complete.

active - The Capacity Reservation Fleet has fulfilled its total target capacity and it is attempting to maintain this capacity. The Fleet remains in this state until it is modified or deleted.

partially_fulfilled - The Capacity Reservation Fleet has partially fulfilled its total target capacity. There is insufficient Amazon EC2 to fulfill the total target capacity. The Fleet is attempting to asynchronously fulfill its total target capacity.

expiring - The Capacity Reservation Fleet has reach its end date and it is in the process of expiring. One or more of its Capacity reservations might still be active.

expired - The Capacity Reservation Fleet has reach its end date. The Fleet and its Capacity Reservations are expired. The Fleet can't create new Capacity Reservations.

cancelling - The Capacity Reservation Fleet is in the process of being cancelled. One or more of its Capacity reservations might still be active.

cancelled - The Capacity Reservation Fleet has been manually cancelled. The Fleet and its Capacity Reservations are cancelled and the Fleet can't create new Capacity Reservations.

failed - The Capacity Reservation Fleet failed to reserve capacity for the specified instance types.

capacityReservationFleet_totalFulfilledCapacity :: Lens' CapacityReservationFleet (Maybe Double)

The capacity units that have been fulfilled.

capacityReservationFleet_endDate :: Lens' CapacityReservationFleet (Maybe UTCTime)

The date and time at which the Capacity Reservation Fleet expires.

capacityReservationFleet_instanceMatchCriteria :: Lens' CapacityReservationFleet (Maybe FleetInstanceMatchCriteria)

Indicates the type of instance launches that the Capacity Reservation Fleet accepts. All Capacity Reservations in the Fleet inherit this instance matching criteria.

Currently, Capacity Reservation Fleets support open instance matching criteria only. This means that instances that have matching attributes (instance type, platform, and Availability Zone) run in the Capacity Reservations automatically. Instances do not need to explicitly target a Capacity Reservation Fleet to use its reserved capacity.

capacityReservationFleet_totalTargetCapacity :: Lens' CapacityReservationFleet (Maybe Int)

The total number of capacity units for which the Capacity Reservation Fleet reserves capacity. For more information, see Total target capacity in the Amazon EC2 User Guide.

capacityReservationFleet_capacityReservationFleetArn :: Lens' CapacityReservationFleet (Maybe Text)

The ARN of the Capacity Reservation Fleet.

capacityReservationFleet_instanceTypeSpecifications :: Lens' CapacityReservationFleet (Maybe [FleetCapacityReservation])

Information about the instance types for which to reserve the capacity.

capacityReservationFleet_tenancy :: Lens' CapacityReservationFleet (Maybe FleetCapacityReservationTenancy)

The tenancy of the Capacity Reservation Fleet. Tenancies include:

default - The Capacity Reservation Fleet is created on hardware that is shared with other Amazon Web Services accounts.

dedicated - The Capacity Reservation Fleet is created on single-tenant hardware that is dedicated to a single Amazon Web Services account.

capacityReservationFleet_allocationStrategy :: Lens' CapacityReservationFleet (Maybe Text)

The strategy used by the Capacity Reservation Fleet to determine which of the specified instance types to use. For more information, see For more information, see Allocation strategy in the Amazon EC2 User Guide.

capacityReservationFleet_createTime :: Lens' CapacityReservationFleet (Maybe UTCTime)

The date and time at which the Capacity Reservation Fleet was created.

capacityReservationFleet_tags :: Lens' CapacityReservationFleet (Maybe [Tag])

The tags assigned to the Capacity Reservation Fleet.