Amazonka.EC2.Types.ClientConnectResponseOptions

Description

Documentation

data ClientConnectResponseOptions

The options for managing connection authorization for new client connections.

See: newClientConnectResponseOptions smart constructor.

Constructors

ClientConnectResponseOptions'

Fields

status :: Maybe ClientVpnEndpointAttributeStatus

The status of any updates to the client connect options.

enabled :: Maybe Bool

Indicates whether client connect options are enabled.

lambdaFunctionArn :: Maybe Text

The Amazon Resource Name (ARN) of the Lambda function used for connection authorization.

newClientConnectResponseOptions :: ClientConnectResponseOptions

Create a value of ClientConnectResponseOptions with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:status:ClientConnectResponseOptions', clientConnectResponseOptions_status - The status of any updates to the client connect options.

$sel:enabled:ClientConnectResponseOptions', clientConnectResponseOptions_enabled - Indicates whether client connect options are enabled.

$sel:lambdaFunctionArn:ClientConnectResponseOptions', clientConnectResponseOptions_lambdaFunctionArn - The Amazon Resource Name (ARN) of the Lambda function used for connection authorization.

clientConnectResponseOptions_status :: Lens' ClientConnectResponseOptions (Maybe ClientVpnEndpointAttributeStatus)

The status of any updates to the client connect options.

clientConnectResponseOptions_enabled :: Lens' ClientConnectResponseOptions (Maybe Bool)

Indicates whether client connect options are enabled.

clientConnectResponseOptions_lambdaFunctionArn :: Lens' ClientConnectResponseOptions (Maybe Text)

The Amazon Resource Name (ARN) of the Lambda function used for connection authorization.