Amazonka.EC2.Types.ClientVpnAuthentication

Description

Documentation

data ClientVpnAuthentication

Describes the authentication methods used by a Client VPN endpoint. For more information, see Authentication in the Client VPN Administrator Guide.

See: newClientVpnAuthentication smart constructor.

Constructors

ClientVpnAuthentication'

Fields

activeDirectory :: Maybe DirectoryServiceAuthentication

Information about the Active Directory, if applicable.

federatedAuthentication :: Maybe FederatedAuthentication

Information about the IAM SAML identity provider, if applicable.

mutualAuthentication :: Maybe CertificateAuthentication

Information about the authentication certificates, if applicable.

type' :: Maybe ClientVpnAuthenticationType

The authentication type used.

newClientVpnAuthentication :: ClientVpnAuthentication

Create a value of ClientVpnAuthentication with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:activeDirectory:ClientVpnAuthentication', clientVpnAuthentication_activeDirectory - Information about the Active Directory, if applicable.

$sel:federatedAuthentication:ClientVpnAuthentication', clientVpnAuthentication_federatedAuthentication - Information about the IAM SAML identity provider, if applicable.

$sel:mutualAuthentication:ClientVpnAuthentication', clientVpnAuthentication_mutualAuthentication - Information about the authentication certificates, if applicable.

$sel:type':ClientVpnAuthentication', clientVpnAuthentication_type - The authentication type used.

clientVpnAuthentication_activeDirectory :: Lens' ClientVpnAuthentication (Maybe DirectoryServiceAuthentication)

Information about the Active Directory, if applicable.

clientVpnAuthentication_federatedAuthentication :: Lens' ClientVpnAuthentication (Maybe FederatedAuthentication)

Information about the IAM SAML identity provider, if applicable.

clientVpnAuthentication_mutualAuthentication :: Lens' ClientVpnAuthentication (Maybe CertificateAuthentication)

Information about the authentication certificates, if applicable.

clientVpnAuthentication_type :: Lens' ClientVpnAuthentication (Maybe ClientVpnAuthenticationType)

The authentication type used.