Amazonka.EC2.Types.ConnectionLogOptions

Description

Documentation

data ConnectionLogOptions

Describes the client connection logging options for the Client VPN endpoint.

See: newConnectionLogOptions smart constructor.

Constructors

ConnectionLogOptions'

Fields

enabled :: Maybe Bool

Indicates whether connection logging is enabled.

cloudwatchLogStream :: Maybe Text

The name of the CloudWatch Logs log stream to which the connection data is published.

cloudwatchLogGroup :: Maybe Text

The name of the CloudWatch Logs log group. Required if connection logging is enabled.

newConnectionLogOptions :: ConnectionLogOptions

Create a value of ConnectionLogOptions with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:enabled:ConnectionLogOptions', connectionLogOptions_enabled - Indicates whether connection logging is enabled.

$sel:cloudwatchLogStream:ConnectionLogOptions', connectionLogOptions_cloudwatchLogStream - The name of the CloudWatch Logs log stream to which the connection data is published.

$sel:cloudwatchLogGroup:ConnectionLogOptions', connectionLogOptions_cloudwatchLogGroup - The name of the CloudWatch Logs log group. Required if connection logging is enabled.

connectionLogOptions_enabled :: Lens' ConnectionLogOptions (Maybe Bool)

Indicates whether connection logging is enabled.

connectionLogOptions_cloudwatchLogStream :: Lens' ConnectionLogOptions (Maybe Text)

The name of the CloudWatch Logs log stream to which the connection data is published.

connectionLogOptions_cloudwatchLogGroup :: Lens' ConnectionLogOptions (Maybe Text)

The name of the CloudWatch Logs log group. Required if connection logging is enabled.