Amazonka.EC2.Types.DeleteLaunchTemplateVersionsResponseErrorItem

Description

Documentation

data DeleteLaunchTemplateVersionsResponseErrorItem

Describes a launch template version that could not be deleted.

See: newDeleteLaunchTemplateVersionsResponseErrorItem smart constructor.

Constructors

DeleteLaunchTemplateVersionsResponseErrorItem'

Fields

launchTemplateName :: Maybe Text

The name of the launch template.

launchTemplateId :: Maybe Text

The ID of the launch template.

versionNumber :: Maybe Integer

The version number of the launch template.

responseError :: Maybe ResponseError

Information about the error.

newDeleteLaunchTemplateVersionsResponseErrorItem :: DeleteLaunchTemplateVersionsResponseErrorItem

Create a value of DeleteLaunchTemplateVersionsResponseErrorItem with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:launchTemplateName:DeleteLaunchTemplateVersionsResponseErrorItem', deleteLaunchTemplateVersionsResponseErrorItem_launchTemplateName - The name of the launch template.

$sel:launchTemplateId:DeleteLaunchTemplateVersionsResponseErrorItem', deleteLaunchTemplateVersionsResponseErrorItem_launchTemplateId - The ID of the launch template.

$sel:versionNumber:DeleteLaunchTemplateVersionsResponseErrorItem', deleteLaunchTemplateVersionsResponseErrorItem_versionNumber - The version number of the launch template.

$sel:responseError:DeleteLaunchTemplateVersionsResponseErrorItem', deleteLaunchTemplateVersionsResponseErrorItem_responseError - Information about the error.

deleteLaunchTemplateVersionsResponseErrorItem_launchTemplateName :: Lens' DeleteLaunchTemplateVersionsResponseErrorItem (Maybe Text)

The name of the launch template.

deleteLaunchTemplateVersionsResponseErrorItem_launchTemplateId :: Lens' DeleteLaunchTemplateVersionsResponseErrorItem (Maybe Text)

The ID of the launch template.

deleteLaunchTemplateVersionsResponseErrorItem_versionNumber :: Lens' DeleteLaunchTemplateVersionsResponseErrorItem (Maybe Integer)

The version number of the launch template.

deleteLaunchTemplateVersionsResponseErrorItem_responseError :: Lens' DeleteLaunchTemplateVersionsResponseErrorItem (Maybe ResponseError)

Information about the error.