Amazonka.EC2.Types.DeleteLaunchTemplateVersionsResponseSuccessItem

Description

Documentation

data DeleteLaunchTemplateVersionsResponseSuccessItem

Describes a launch template version that was successfully deleted.

See: newDeleteLaunchTemplateVersionsResponseSuccessItem smart constructor.

Constructors

DeleteLaunchTemplateVersionsResponseSuccessItem'

Fields

launchTemplateName :: Maybe Text

The name of the launch template.

launchTemplateId :: Maybe Text

The ID of the launch template.

versionNumber :: Maybe Integer

The version number of the launch template.

newDeleteLaunchTemplateVersionsResponseSuccessItem :: DeleteLaunchTemplateVersionsResponseSuccessItem

Create a value of DeleteLaunchTemplateVersionsResponseSuccessItem with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:launchTemplateName:DeleteLaunchTemplateVersionsResponseSuccessItem', deleteLaunchTemplateVersionsResponseSuccessItem_launchTemplateName - The name of the launch template.

$sel:launchTemplateId:DeleteLaunchTemplateVersionsResponseSuccessItem', deleteLaunchTemplateVersionsResponseSuccessItem_launchTemplateId - The ID of the launch template.

$sel:versionNumber:DeleteLaunchTemplateVersionsResponseSuccessItem', deleteLaunchTemplateVersionsResponseSuccessItem_versionNumber - The version number of the launch template.

deleteLaunchTemplateVersionsResponseSuccessItem_launchTemplateName :: Lens' DeleteLaunchTemplateVersionsResponseSuccessItem (Maybe Text)

The name of the launch template.

deleteLaunchTemplateVersionsResponseSuccessItem_launchTemplateId :: Lens' DeleteLaunchTemplateVersionsResponseSuccessItem (Maybe Text)

The ID of the launch template.

deleteLaunchTemplateVersionsResponseSuccessItem_versionNumber :: Lens' DeleteLaunchTemplateVersionsResponseSuccessItem (Maybe Integer)

The version number of the launch template.