Amazonka.EC2.Types.DhcpOptions

Description

Documentation

data DhcpOptions

Describes a set of DHCP options.

See: newDhcpOptions smart constructor.

Constructors

DhcpOptions'

Fields

dhcpConfigurations :: Maybe [DhcpConfiguration]

One or more DHCP options in the set.

ownerId :: Maybe Text

The ID of the Amazon Web Services account that owns the DHCP options set.

dhcpOptionsId :: Maybe Text

The ID of the set of DHCP options.

tags :: Maybe [Tag]

Any tags assigned to the DHCP options set.

newDhcpOptions :: DhcpOptions

Create a value of DhcpOptions with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:dhcpConfigurations:DhcpOptions', dhcpOptions_dhcpConfigurations - One or more DHCP options in the set.

$sel:ownerId:DhcpOptions', dhcpOptions_ownerId - The ID of the Amazon Web Services account that owns the DHCP options set.

$sel:dhcpOptionsId:DhcpOptions', dhcpOptions_dhcpOptionsId - The ID of the set of DHCP options.

$sel:tags:DhcpOptions', dhcpOptions_tags - Any tags assigned to the DHCP options set.

dhcpOptions_dhcpConfigurations :: Lens' DhcpOptions (Maybe [DhcpConfiguration])

One or more DHCP options in the set.

dhcpOptions_ownerId :: Lens' DhcpOptions (Maybe Text)

The ID of the Amazon Web Services account that owns the DHCP options set.

dhcpOptions_dhcpOptionsId :: Lens' DhcpOptions (Maybe Text)

The ID of the set of DHCP options.

dhcpOptions_tags :: Lens' DhcpOptions (Maybe [Tag])

Any tags assigned to the DHCP options set.