Amazonka.EC2.Types.DiskImageDescription

Description

Documentation

data DiskImageDescription

Describes a disk image.

See: newDiskImageDescription smart constructor.

Constructors

DiskImageDescription'

Fields

size :: Maybe Integer

The size of the disk image, in GiB.

checksum :: Maybe Text

The checksum computed for the disk image.

format :: Maybe DiskImageFormat

The disk image format.

importManifestUrl :: Maybe Text

A presigned URL for the import manifest stored in Amazon S3. For information about creating a presigned URL for an Amazon S3 object, read the "Query String Request Authentication Alternative" section of the Authenticating REST Requests topic in the Amazon Simple Storage Service Developer Guide.

For information about the import manifest referenced by this API action, see VM Import Manifest.

newDiskImageDescription :: DiskImageDescription

Create a value of DiskImageDescription with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:size:DiskImageDescription', diskImageDescription_size - The size of the disk image, in GiB.

$sel:checksum:DiskImageDescription', diskImageDescription_checksum - The checksum computed for the disk image.

$sel:format:DiskImageDescription', diskImageDescription_format - The disk image format.

$sel:importManifestUrl:DiskImageDescription', diskImageDescription_importManifestUrl - A presigned URL for the import manifest stored in Amazon S3. For information about creating a presigned URL for an Amazon S3 object, read the "Query String Request Authentication Alternative" section of the Authenticating REST Requests topic in the Amazon Simple Storage Service Developer Guide.

For information about the import manifest referenced by this API action, see VM Import Manifest.

diskImageDescription_size :: Lens' DiskImageDescription (Maybe Integer)

The size of the disk image, in GiB.

diskImageDescription_checksum :: Lens' DiskImageDescription (Maybe Text)

The checksum computed for the disk image.

diskImageDescription_format :: Lens' DiskImageDescription (Maybe DiskImageFormat)

The disk image format.

diskImageDescription_importManifestUrl :: Lens' DiskImageDescription (Maybe Text)

A presigned URL for the import manifest stored in Amazon S3. For information about creating a presigned URL for an Amazon S3 object, read the "Query String Request Authentication Alternative" section of the Authenticating REST Requests topic in the Amazon Simple Storage Service Developer Guide.

For information about the import manifest referenced by this API action, see VM Import Manifest.