Amazonka.EC2.Types.Explanation

Description

Documentation

data Explanation

Describes an explanation code for an unreachable path. For more information, see Reachability Analyzer explanation codes.

See: newExplanation smart constructor.

Constructors

Explanation'

Fields

destination :: Maybe AnalysisComponent

The destination.

state :: Maybe Text

The state.

cidrs :: Maybe [Text]

The CIDR ranges.

component :: Maybe AnalysisComponent

The component.

loadBalancerTargetGroups :: Maybe [AnalysisComponent]

The target groups.

securityGroups :: Maybe [AnalysisComponent]

The security groups.

prefixList :: Maybe AnalysisComponent

The prefix list.

direction :: Maybe Text

The direction. The following are possible values:

egress

ingress

protocols :: Maybe [Text]

The protocols.

loadBalancerListenerPort :: Maybe Natural

The listener port of the load balancer.

portRanges :: Maybe [PortRange]

The port ranges.

addresses :: Maybe [Text]

The IPv4 addresses, in CIDR notation.

classicLoadBalancerListener :: Maybe AnalysisLoadBalancerListener

The listener for a Classic Load Balancer.

ingressRouteTable :: Maybe AnalysisComponent

The route table.

networkInterface :: Maybe AnalysisComponent

The network interface.

loadBalancerTarget :: Maybe AnalysisLoadBalancerTarget

The target.

subnet :: Maybe AnalysisComponent

The subnet.

natGateway :: Maybe AnalysisComponent

The NAT gateway.

address :: Maybe Text

The IPv4 address, in CIDR notation.

explanationCode :: Maybe Text

The explanation code.

securityGroup :: Maybe AnalysisComponent

The security group.

elasticLoadBalancerListener :: Maybe AnalysisComponent

The load balancer listener.

loadBalancerTargetGroup :: Maybe AnalysisComponent

The target group.

customerGateway :: Maybe AnalysisComponent

The customer gateway.

subnetRouteTable :: Maybe AnalysisComponent

The route table for the subnet.

availabilityZones :: Maybe [Text]

The Availability Zones.

loadBalancerArn :: Maybe Text

The Amazon Resource Name (ARN) of the load balancer.

routeTable :: Maybe AnalysisComponent

The route table.

securityGroupRule :: Maybe AnalysisSecurityGroupRule

The security group rule.

packetField :: Maybe Text

The packet field.

loadBalancerTargetPort :: Maybe Natural

The target port.

vpc :: Maybe AnalysisComponent

The component VPC.

vpnGateway :: Maybe AnalysisComponent

The VPN gateway.

sourceVpc :: Maybe AnalysisComponent

The source VPC.

aclRule :: Maybe AnalysisAclRule

The network ACL rule.

internetGateway :: Maybe AnalysisComponent

The internet gateway.

missingComponent :: Maybe Text

The missing component.

acl :: Maybe AnalysisComponent

The network ACL.

vpnConnection :: Maybe AnalysisComponent

The VPN connection.

routeTableRoute :: Maybe AnalysisRouteTableRoute

The route table route.

vpcEndpoint :: Maybe AnalysisComponent

The VPC endpoint.

vpcPeeringConnection :: Maybe AnalysisComponent

The VPC peering connection.

port :: Maybe Natural

The port.

destinationVpc :: Maybe AnalysisComponent

The destination VPC.

attachedTo :: Maybe AnalysisComponent

The resource to which the component is attached.

newExplanation :: Explanation

Create a value of Explanation with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:destination:Explanation', explanation_destination - The destination.

$sel:state:Explanation', explanation_state - The state.

$sel:cidrs:Explanation', explanation_cidrs - The CIDR ranges.

$sel:component:Explanation', explanation_component - The component.

$sel:loadBalancerTargetGroups:Explanation', explanation_loadBalancerTargetGroups - The target groups.

$sel:securityGroups:Explanation', explanation_securityGroups - The security groups.

$sel:prefixList:Explanation', explanation_prefixList - The prefix list.

$sel:direction:Explanation', explanation_direction - The direction. The following are possible values:

egress

ingress

$sel:protocols:Explanation', explanation_protocols - The protocols.

$sel:loadBalancerListenerPort:Explanation', explanation_loadBalancerListenerPort - The listener port of the load balancer.

$sel:portRanges:Explanation', explanation_portRanges - The port ranges.

$sel:addresses:Explanation', explanation_addresses - The IPv4 addresses, in CIDR notation.

$sel:classicLoadBalancerListener:Explanation', explanation_classicLoadBalancerListener - The listener for a Classic Load Balancer.

$sel:ingressRouteTable:Explanation', explanation_ingressRouteTable - The route table.

$sel:networkInterface:Explanation', explanation_networkInterface - The network interface.

$sel:loadBalancerTarget:Explanation', explanation_loadBalancerTarget - The target.

$sel:subnet:Explanation', explanation_subnet - The subnet.

$sel:natGateway:Explanation', explanation_natGateway - The NAT gateway.

$sel:address:Explanation', explanation_address - The IPv4 address, in CIDR notation.

$sel:explanationCode:Explanation', explanation_explanationCode - The explanation code.

$sel:securityGroup:Explanation', explanation_securityGroup - The security group.

$sel:elasticLoadBalancerListener:Explanation', explanation_elasticLoadBalancerListener - The load balancer listener.

$sel:loadBalancerTargetGroup:Explanation', explanation_loadBalancerTargetGroup - The target group.

$sel:customerGateway:Explanation', explanation_customerGateway - The customer gateway.

$sel:subnetRouteTable:Explanation', explanation_subnetRouteTable - The route table for the subnet.

$sel:availabilityZones:Explanation', explanation_availabilityZones - The Availability Zones.

$sel:loadBalancerArn:Explanation', explanation_loadBalancerArn - The Amazon Resource Name (ARN) of the load balancer.

$sel:routeTable:Explanation', explanation_routeTable - The route table.

$sel:securityGroupRule:Explanation', explanation_securityGroupRule - The security group rule.

$sel:packetField:Explanation', explanation_packetField - The packet field.

$sel:loadBalancerTargetPort:Explanation', explanation_loadBalancerTargetPort - The target port.

$sel:vpc:Explanation', explanation_vpc - The component VPC.

$sel:vpnGateway:Explanation', explanation_vpnGateway - The VPN gateway.

$sel:sourceVpc:Explanation', explanation_sourceVpc - The source VPC.

$sel:aclRule:Explanation', explanation_aclRule - The network ACL rule.

$sel:internetGateway:Explanation', explanation_internetGateway - The internet gateway.

$sel:missingComponent:Explanation', explanation_missingComponent - The missing component.

$sel:acl:Explanation', explanation_acl - The network ACL.

$sel:vpnConnection:Explanation', explanation_vpnConnection - The VPN connection.

$sel:routeTableRoute:Explanation', explanation_routeTableRoute - The route table route.

$sel:vpcEndpoint:Explanation', explanation_vpcEndpoint - The VPC endpoint.

$sel:vpcPeeringConnection:Explanation', explanation_vpcPeeringConnection - The VPC peering connection.

$sel:port:Explanation', explanation_port - The port.

$sel:destinationVpc:Explanation', explanation_destinationVpc - The destination VPC.

$sel:attachedTo:Explanation', explanation_attachedTo - The resource to which the component is attached.

explanation_destination :: Lens' Explanation (Maybe AnalysisComponent)

The destination.

explanation_state :: Lens' Explanation (Maybe Text)

The state.

explanation_cidrs :: Lens' Explanation (Maybe [Text])

The CIDR ranges.

explanation_component :: Lens' Explanation (Maybe AnalysisComponent)

The component.

explanation_loadBalancerTargetGroups :: Lens' Explanation (Maybe [AnalysisComponent])

The target groups.

explanation_securityGroups :: Lens' Explanation (Maybe [AnalysisComponent])

The security groups.

explanation_prefixList :: Lens' Explanation (Maybe AnalysisComponent)

The prefix list.

explanation_direction :: Lens' Explanation (Maybe Text)

The direction. The following are possible values:

egress

ingress

explanation_protocols :: Lens' Explanation (Maybe [Text])

The protocols.

explanation_loadBalancerListenerPort :: Lens' Explanation (Maybe Natural)

The listener port of the load balancer.

explanation_portRanges :: Lens' Explanation (Maybe [PortRange])

The port ranges.

explanation_addresses :: Lens' Explanation (Maybe [Text])

The IPv4 addresses, in CIDR notation.

explanation_classicLoadBalancerListener :: Lens' Explanation (Maybe AnalysisLoadBalancerListener)

The listener for a Classic Load Balancer.

explanation_ingressRouteTable :: Lens' Explanation (Maybe AnalysisComponent)

The route table.

explanation_networkInterface :: Lens' Explanation (Maybe AnalysisComponent)

The network interface.

explanation_loadBalancerTarget :: Lens' Explanation (Maybe AnalysisLoadBalancerTarget)

The target.

explanation_subnet :: Lens' Explanation (Maybe AnalysisComponent)

The subnet.

explanation_natGateway :: Lens' Explanation (Maybe AnalysisComponent)

The NAT gateway.

explanation_address :: Lens' Explanation (Maybe Text)

The IPv4 address, in CIDR notation.

explanation_explanationCode :: Lens' Explanation (Maybe Text)

The explanation code.

explanation_securityGroup :: Lens' Explanation (Maybe AnalysisComponent)

The security group.

explanation_elasticLoadBalancerListener :: Lens' Explanation (Maybe AnalysisComponent)

The load balancer listener.

explanation_loadBalancerTargetGroup :: Lens' Explanation (Maybe AnalysisComponent)

The target group.

explanation_customerGateway :: Lens' Explanation (Maybe AnalysisComponent)

The customer gateway.

explanation_subnetRouteTable :: Lens' Explanation (Maybe AnalysisComponent)

The route table for the subnet.

explanation_availabilityZones :: Lens' Explanation (Maybe [Text])

The Availability Zones.

explanation_loadBalancerArn :: Lens' Explanation (Maybe Text)

The Amazon Resource Name (ARN) of the load balancer.

explanation_routeTable :: Lens' Explanation (Maybe AnalysisComponent)

The route table.

explanation_securityGroupRule :: Lens' Explanation (Maybe AnalysisSecurityGroupRule)

The security group rule.

explanation_packetField :: Lens' Explanation (Maybe Text)

The packet field.

explanation_loadBalancerTargetPort :: Lens' Explanation (Maybe Natural)

The target port.

explanation_vpc :: Lens' Explanation (Maybe AnalysisComponent)

The component VPC.

explanation_vpnGateway :: Lens' Explanation (Maybe AnalysisComponent)

The VPN gateway.

explanation_sourceVpc :: Lens' Explanation (Maybe AnalysisComponent)

The source VPC.

explanation_aclRule :: Lens' Explanation (Maybe AnalysisAclRule)

The network ACL rule.

explanation_internetGateway :: Lens' Explanation (Maybe AnalysisComponent)

The internet gateway.

explanation_missingComponent :: Lens' Explanation (Maybe Text)

The missing component.

explanation_acl :: Lens' Explanation (Maybe AnalysisComponent)

The network ACL.

explanation_vpnConnection :: Lens' Explanation (Maybe AnalysisComponent)

The VPN connection.

explanation_routeTableRoute :: Lens' Explanation (Maybe AnalysisRouteTableRoute)

The route table route.

explanation_vpcEndpoint :: Lens' Explanation (Maybe AnalysisComponent)

The VPC endpoint.

explanation_vpcPeeringConnection :: Lens' Explanation (Maybe AnalysisComponent)

The VPC peering connection.

explanation_port :: Lens' Explanation (Maybe Natural)

The port.

explanation_destinationVpc :: Lens' Explanation (Maybe AnalysisComponent)

The destination VPC.

explanation_attachedTo :: Lens' Explanation (Maybe AnalysisComponent)

The resource to which the component is attached.