Amazonka.EC2.Types.ExportToS3TaskSpecification

Description

Documentation

data ExportToS3TaskSpecification

Describes an export instance task.

See: newExportToS3TaskSpecification smart constructor.

Constructors

ExportToS3TaskSpecification'

Fields

containerFormat :: Maybe ContainerFormat

The container format used to combine disk images with metadata (such as OVF). If absent, only the disk image is exported.

s3Prefix :: Maybe Text

The image is written to a single object in the Amazon S3 bucket at the S3 key s3prefix + exportTaskId + '.' + diskImageFormat.

s3Bucket :: Maybe Text

The Amazon S3 bucket for the destination image. The destination bucket must exist and grant WRITE and READ_ACP permissions to the Amazon Web Services account vm-import-export@amazon.com.

diskImageFormat :: Maybe DiskImageFormat

The format for the exported image.

newExportToS3TaskSpecification :: ExportToS3TaskSpecification

Create a value of ExportToS3TaskSpecification with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:containerFormat:ExportToS3TaskSpecification', exportToS3TaskSpecification_containerFormat - The container format used to combine disk images with metadata (such as OVF). If absent, only the disk image is exported.

$sel:s3Prefix:ExportToS3TaskSpecification', exportToS3TaskSpecification_s3Prefix - The image is written to a single object in the Amazon S3 bucket at the S3 key s3prefix + exportTaskId + '.' + diskImageFormat.

$sel:s3Bucket:ExportToS3TaskSpecification', exportToS3TaskSpecification_s3Bucket - The Amazon S3 bucket for the destination image. The destination bucket must exist and grant WRITE and READ_ACP permissions to the Amazon Web Services account vm-import-export@amazon.com.

$sel:diskImageFormat:ExportToS3TaskSpecification', exportToS3TaskSpecification_diskImageFormat - The format for the exported image.

exportToS3TaskSpecification_containerFormat :: Lens' ExportToS3TaskSpecification (Maybe ContainerFormat)

The container format used to combine disk images with metadata (such as OVF). If absent, only the disk image is exported.

exportToS3TaskSpecification_s3Prefix :: Lens' ExportToS3TaskSpecification (Maybe Text)

The image is written to a single object in the Amazon S3 bucket at the S3 key s3prefix + exportTaskId + '.' + diskImageFormat.

exportToS3TaskSpecification_s3Bucket :: Lens' ExportToS3TaskSpecification (Maybe Text)

The Amazon S3 bucket for the destination image. The destination bucket must exist and grant WRITE and READ_ACP permissions to the Amazon Web Services account vm-import-export@amazon.com.

exportToS3TaskSpecification_diskImageFormat :: Lens' ExportToS3TaskSpecification (Maybe DiskImageFormat)

The format for the exported image.