Amazonka.EC2.Types.FlowLog

Description

Documentation

data FlowLog

Describes a flow log.

See: newFlowLog smart constructor.

Constructors

FlowLog'

Fields

creationTime :: Maybe ISO8601

The date and time the flow log was created.

logFormat :: Maybe Text

The format of the flow log record.

maxAggregationInterval :: Maybe Int

The maximum interval of time, in seconds, during which a flow of packets is captured and aggregated into a flow log record.

When a network interface is attached to a Nitro-based instance, the aggregation interval is always 60 seconds (1 minute) or less, regardless of the specified value.

Valid Values: 60 | 600

resourceId :: Maybe Text

The ID of the resource on which the flow log was created.

flowLogStatus :: Maybe Text

The status of the flow log (ACTIVE).

trafficType :: Maybe TrafficType

The type of traffic captured for the flow log.

logDestination :: Maybe Text

The destination to which the flow log data is published. Flow log data can be published to an CloudWatch Logs log group or an Amazon S3 bucket. If the flow log publishes to CloudWatch Logs, this element indicates the Amazon Resource Name (ARN) of the CloudWatch Logs log group to which the data is published. If the flow log publishes to Amazon S3, this element indicates the ARN of the Amazon S3 bucket to which the data is published.

deliverLogsStatus :: Maybe Text

The status of the logs delivery (SUCCESS | FAILED).

deliverLogsErrorMessage :: Maybe Text

Information about the error that occurred. Rate limited indicates that CloudWatch Logs throttling has been applied for one or more network interfaces, or that you've reached the limit on the number of log groups that you can create. Access error indicates that the IAM role associated with the flow log does not have sufficient permissions to publish to CloudWatch Logs. Unknown error indicates an internal error.

logGroupName :: Maybe Text

The name of the flow log group.

destinationOptions :: Maybe DestinationOptionsResponse

The destination options.

deliverLogsPermissionArn :: Maybe Text

The ARN of the IAM role that posts logs to CloudWatch Logs.

logDestinationType :: Maybe LogDestinationType

The type of destination to which the flow log data is published. Flow log data can be published to CloudWatch Logs or Amazon S3.

flowLogId :: Maybe Text

The flow log ID.

tags :: Maybe [Tag]

The tags for the flow log.

newFlowLog :: FlowLog

Create a value of FlowLog with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:creationTime:FlowLog', flowLog_creationTime - The date and time the flow log was created.

$sel:logFormat:FlowLog', flowLog_logFormat - The format of the flow log record.

$sel:maxAggregationInterval:FlowLog', flowLog_maxAggregationInterval - The maximum interval of time, in seconds, during which a flow of packets is captured and aggregated into a flow log record.

When a network interface is attached to a Nitro-based instance, the aggregation interval is always 60 seconds (1 minute) or less, regardless of the specified value.

Valid Values: 60 | 600

$sel:resourceId:FlowLog', flowLog_resourceId - The ID of the resource on which the flow log was created.

$sel:flowLogStatus:FlowLog', flowLog_flowLogStatus - The status of the flow log (ACTIVE).

$sel:trafficType:FlowLog', flowLog_trafficType - The type of traffic captured for the flow log.

$sel:logDestination:FlowLog', flowLog_logDestination - The destination to which the flow log data is published. Flow log data can be published to an CloudWatch Logs log group or an Amazon S3 bucket. If the flow log publishes to CloudWatch Logs, this element indicates the Amazon Resource Name (ARN) of the CloudWatch Logs log group to which the data is published. If the flow log publishes to Amazon S3, this element indicates the ARN of the Amazon S3 bucket to which the data is published.

$sel:deliverLogsStatus:FlowLog', flowLog_deliverLogsStatus - The status of the logs delivery (SUCCESS | FAILED).

$sel:deliverLogsErrorMessage:FlowLog', flowLog_deliverLogsErrorMessage - Information about the error that occurred. Rate limited indicates that CloudWatch Logs throttling has been applied for one or more network interfaces, or that you've reached the limit on the number of log groups that you can create. Access error indicates that the IAM role associated with the flow log does not have sufficient permissions to publish to CloudWatch Logs. Unknown error indicates an internal error.

$sel:logGroupName:FlowLog', flowLog_logGroupName - The name of the flow log group.

$sel:destinationOptions:FlowLog', flowLog_destinationOptions - The destination options.

$sel:deliverLogsPermissionArn:FlowLog', flowLog_deliverLogsPermissionArn - The ARN of the IAM role that posts logs to CloudWatch Logs.

$sel:logDestinationType:FlowLog', flowLog_logDestinationType - The type of destination to which the flow log data is published. Flow log data can be published to CloudWatch Logs or Amazon S3.

$sel:flowLogId:FlowLog', flowLog_flowLogId - The flow log ID.

$sel:tags:FlowLog', flowLog_tags - The tags for the flow log.

flowLog_creationTime :: Lens' FlowLog (Maybe UTCTime)

The date and time the flow log was created.

flowLog_logFormat :: Lens' FlowLog (Maybe Text)

The format of the flow log record.

flowLog_maxAggregationInterval :: Lens' FlowLog (Maybe Int)

The maximum interval of time, in seconds, during which a flow of packets is captured and aggregated into a flow log record.

When a network interface is attached to a Nitro-based instance, the aggregation interval is always 60 seconds (1 minute) or less, regardless of the specified value.

Valid Values: 60 | 600

flowLog_resourceId :: Lens' FlowLog (Maybe Text)

The ID of the resource on which the flow log was created.

flowLog_flowLogStatus :: Lens' FlowLog (Maybe Text)

The status of the flow log (ACTIVE).

flowLog_trafficType :: Lens' FlowLog (Maybe TrafficType)

The type of traffic captured for the flow log.

flowLog_logDestination :: Lens' FlowLog (Maybe Text)

The destination to which the flow log data is published. Flow log data can be published to an CloudWatch Logs log group or an Amazon S3 bucket. If the flow log publishes to CloudWatch Logs, this element indicates the Amazon Resource Name (ARN) of the CloudWatch Logs log group to which the data is published. If the flow log publishes to Amazon S3, this element indicates the ARN of the Amazon S3 bucket to which the data is published.

flowLog_deliverLogsStatus :: Lens' FlowLog (Maybe Text)

The status of the logs delivery (SUCCESS | FAILED).

flowLog_deliverLogsErrorMessage :: Lens' FlowLog (Maybe Text)

Information about the error that occurred. Rate limited indicates that CloudWatch Logs throttling has been applied for one or more network interfaces, or that you've reached the limit on the number of log groups that you can create. Access error indicates that the IAM role associated with the flow log does not have sufficient permissions to publish to CloudWatch Logs. Unknown error indicates an internal error.

flowLog_logGroupName :: Lens' FlowLog (Maybe Text)

The name of the flow log group.

flowLog_destinationOptions :: Lens' FlowLog (Maybe DestinationOptionsResponse)

The destination options.

flowLog_deliverLogsPermissionArn :: Lens' FlowLog (Maybe Text)

The ARN of the IAM role that posts logs to CloudWatch Logs.

flowLog_logDestinationType :: Lens' FlowLog (Maybe LogDestinationType)

The type of destination to which the flow log data is published. Flow log data can be published to CloudWatch Logs or Amazon S3.

flowLog_flowLogId :: Lens' FlowLog (Maybe Text)

The flow log ID.

flowLog_tags :: Lens' FlowLog (Maybe [Tag])

The tags for the flow log.