Amazonka.EC2.Types.Host

Description

Documentation

data Host

Describes the properties of the Dedicated Host.

See: newHost smart constructor.

Constructors

Host'

Fields

releaseTime :: Maybe ISO8601

The time that the Dedicated Host was released.

state :: Maybe AllocationState

The Dedicated Host's state.

clientToken :: Maybe Text

Unique, case-sensitive identifier that you provide to ensure the idempotency of the request. For more information, see Ensuring Idempotency.

availabilityZoneId :: Maybe Text

The ID of the Availability Zone in which the Dedicated Host is allocated.

hostId :: Maybe Text

The ID of the Dedicated Host.

availableCapacity :: Maybe AvailableCapacity

Information about the instances running on the Dedicated Host.

hostReservationId :: Maybe Text

The reservation ID of the Dedicated Host. This returns a null response if the Dedicated Host doesn't have an associated reservation.

allowsMultipleInstanceTypes :: Maybe AllowsMultipleInstanceTypes

Indicates whether the Dedicated Host supports multiple instance types of the same instance family. If the value is on, the Dedicated Host supports multiple instance types in the instance family. If the value is off, the Dedicated Host supports a single instance type only.

hostProperties :: Maybe HostProperties

The hardware specifications of the Dedicated Host.

ownerId :: Maybe Text

The ID of the Amazon Web Services account that owns the Dedicated Host.

availabilityZone :: Maybe Text

The Availability Zone of the Dedicated Host.

instances :: Maybe [HostInstance]

The IDs and instance type that are currently running on the Dedicated Host.

allocationTime :: Maybe ISO8601

The time that the Dedicated Host was allocated.

memberOfServiceLinkedResourceGroup :: Maybe Bool

Indicates whether the Dedicated Host is in a host resource group. If memberOfServiceLinkedResourceGroup is true, the host is in a host resource group; otherwise, it is not.

hostRecovery :: Maybe HostRecovery

Indicates whether host recovery is enabled or disabled for the Dedicated Host.

autoPlacement :: Maybe AutoPlacement

Whether auto-placement is on or off.

tags :: Maybe [Tag]

Any tags assigned to the Dedicated Host.

newHost :: Host

Create a value of Host with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:releaseTime:Host', host_releaseTime - The time that the Dedicated Host was released.

$sel:state:Host', host_state - The Dedicated Host's state.

$sel:clientToken:Host', host_clientToken - Unique, case-sensitive identifier that you provide to ensure the idempotency of the request. For more information, see Ensuring Idempotency.

$sel:availabilityZoneId:Host', host_availabilityZoneId - The ID of the Availability Zone in which the Dedicated Host is allocated.

$sel:hostId:Host', host_hostId - The ID of the Dedicated Host.

$sel:availableCapacity:Host', host_availableCapacity - Information about the instances running on the Dedicated Host.

$sel:hostReservationId:Host', host_hostReservationId - The reservation ID of the Dedicated Host. This returns a null response if the Dedicated Host doesn't have an associated reservation.

$sel:allowsMultipleInstanceTypes:Host', host_allowsMultipleInstanceTypes - Indicates whether the Dedicated Host supports multiple instance types of the same instance family. If the value is on, the Dedicated Host supports multiple instance types in the instance family. If the value is off, the Dedicated Host supports a single instance type only.

$sel:hostProperties:Host', host_hostProperties - The hardware specifications of the Dedicated Host.

$sel:ownerId:Host', host_ownerId - The ID of the Amazon Web Services account that owns the Dedicated Host.

$sel:availabilityZone:Host', host_availabilityZone - The Availability Zone of the Dedicated Host.

$sel:instances:Host', host_instances - The IDs and instance type that are currently running on the Dedicated Host.

$sel:allocationTime:Host', host_allocationTime - The time that the Dedicated Host was allocated.

$sel:memberOfServiceLinkedResourceGroup:Host', host_memberOfServiceLinkedResourceGroup - Indicates whether the Dedicated Host is in a host resource group. If memberOfServiceLinkedResourceGroup is true, the host is in a host resource group; otherwise, it is not.

$sel:hostRecovery:Host', host_hostRecovery - Indicates whether host recovery is enabled or disabled for the Dedicated Host.

$sel:autoPlacement:Host', host_autoPlacement - Whether auto-placement is on or off.

$sel:tags:Host', host_tags - Any tags assigned to the Dedicated Host.

host_releaseTime :: Lens' Host (Maybe UTCTime)

The time that the Dedicated Host was released.

host_state :: Lens' Host (Maybe AllocationState)

The Dedicated Host's state.

host_clientToken :: Lens' Host (Maybe Text)

Unique, case-sensitive identifier that you provide to ensure the idempotency of the request. For more information, see Ensuring Idempotency.

host_availabilityZoneId :: Lens' Host (Maybe Text)

The ID of the Availability Zone in which the Dedicated Host is allocated.

host_hostId :: Lens' Host (Maybe Text)

The ID of the Dedicated Host.

host_availableCapacity :: Lens' Host (Maybe AvailableCapacity)

Information about the instances running on the Dedicated Host.

host_hostReservationId :: Lens' Host (Maybe Text)

The reservation ID of the Dedicated Host. This returns a null response if the Dedicated Host doesn't have an associated reservation.

host_allowsMultipleInstanceTypes :: Lens' Host (Maybe AllowsMultipleInstanceTypes)

Indicates whether the Dedicated Host supports multiple instance types of the same instance family. If the value is on, the Dedicated Host supports multiple instance types in the instance family. If the value is off, the Dedicated Host supports a single instance type only.

host_hostProperties :: Lens' Host (Maybe HostProperties)

The hardware specifications of the Dedicated Host.

host_ownerId :: Lens' Host (Maybe Text)

The ID of the Amazon Web Services account that owns the Dedicated Host.

host_availabilityZone :: Lens' Host (Maybe Text)

The Availability Zone of the Dedicated Host.

host_instances :: Lens' Host (Maybe [HostInstance])

The IDs and instance type that are currently running on the Dedicated Host.

host_allocationTime :: Lens' Host (Maybe UTCTime)

The time that the Dedicated Host was allocated.

host_memberOfServiceLinkedResourceGroup :: Lens' Host (Maybe Bool)

Indicates whether the Dedicated Host is in a host resource group. If memberOfServiceLinkedResourceGroup is true, the host is in a host resource group; otherwise, it is not.

host_hostRecovery :: Lens' Host (Maybe HostRecovery)

Indicates whether host recovery is enabled or disabled for the Dedicated Host.

host_autoPlacement :: Lens' Host (Maybe AutoPlacement)

Whether auto-placement is on or off.

host_tags :: Lens' Host (Maybe [Tag])

Any tags assigned to the Dedicated Host.