Amazonka.EC2.Types.InstanceEventWindow

Description

Documentation

data InstanceEventWindow

The event window.

See: newInstanceEventWindow smart constructor.

Constructors

InstanceEventWindow'

Fields

state :: Maybe InstanceEventWindowState

The current state of the event window.

associationTarget :: Maybe InstanceEventWindowAssociationTarget

One or more targets associated with the event window.

instanceEventWindowId :: Maybe Text

The ID of the event window.

name :: Maybe Text

The name of the event window.

cronExpression :: Maybe Text

The cron expression defined for the event window.

tags :: Maybe [Tag]

The instance tags associated with the event window.

timeRanges :: Maybe [InstanceEventWindowTimeRange]

One or more time ranges defined for the event window.

newInstanceEventWindow :: InstanceEventWindow

Create a value of InstanceEventWindow with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:state:InstanceEventWindow', instanceEventWindow_state - The current state of the event window.

$sel:associationTarget:InstanceEventWindow', instanceEventWindow_associationTarget - One or more targets associated with the event window.

$sel:instanceEventWindowId:InstanceEventWindow', instanceEventWindow_instanceEventWindowId - The ID of the event window.

$sel:name:InstanceEventWindow', instanceEventWindow_name - The name of the event window.

$sel:cronExpression:InstanceEventWindow', instanceEventWindow_cronExpression - The cron expression defined for the event window.

$sel:tags:InstanceEventWindow', instanceEventWindow_tags - The instance tags associated with the event window.

$sel:timeRanges:InstanceEventWindow', instanceEventWindow_timeRanges - One or more time ranges defined for the event window.

instanceEventWindow_state :: Lens' InstanceEventWindow (Maybe InstanceEventWindowState)

The current state of the event window.

instanceEventWindow_associationTarget :: Lens' InstanceEventWindow (Maybe InstanceEventWindowAssociationTarget)

One or more targets associated with the event window.

instanceEventWindow_instanceEventWindowId :: Lens' InstanceEventWindow (Maybe Text)

The ID of the event window.

instanceEventWindow_name :: Lens' InstanceEventWindow (Maybe Text)

The name of the event window.

instanceEventWindow_cronExpression :: Lens' InstanceEventWindow (Maybe Text)

The cron expression defined for the event window.

instanceEventWindow_tags :: Lens' InstanceEventWindow (Maybe [Tag])

The instance tags associated with the event window.

instanceEventWindow_timeRanges :: Lens' InstanceEventWindow (Maybe [InstanceEventWindowTimeRange])

One or more time ranges defined for the event window.