Amazonka.EC2.Types.InstanceEventWindowTimeRangeRequest

Description

Documentation

data InstanceEventWindowTimeRangeRequest

The start day and time and the end day and time of the time range, in UTC.

See: newInstanceEventWindowTimeRangeRequest smart constructor.

Constructors

InstanceEventWindowTimeRangeRequest'

Fields

endHour :: Maybe Natural

The hour when the time range ends.

endWeekDay :: Maybe WeekDay

The day on which the time range ends.

startWeekDay :: Maybe WeekDay

The day on which the time range begins.

startHour :: Maybe Natural

The hour when the time range begins.

newInstanceEventWindowTimeRangeRequest :: InstanceEventWindowTimeRangeRequest

Create a value of InstanceEventWindowTimeRangeRequest with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:endHour:InstanceEventWindowTimeRangeRequest', instanceEventWindowTimeRangeRequest_endHour - The hour when the time range ends.

$sel:endWeekDay:InstanceEventWindowTimeRangeRequest', instanceEventWindowTimeRangeRequest_endWeekDay - The day on which the time range ends.

$sel:startWeekDay:InstanceEventWindowTimeRangeRequest', instanceEventWindowTimeRangeRequest_startWeekDay - The day on which the time range begins.

$sel:startHour:InstanceEventWindowTimeRangeRequest', instanceEventWindowTimeRangeRequest_startHour - The hour when the time range begins.

instanceEventWindowTimeRangeRequest_endHour :: Lens' InstanceEventWindowTimeRangeRequest (Maybe Natural)

The hour when the time range ends.

instanceEventWindowTimeRangeRequest_endWeekDay :: Lens' InstanceEventWindowTimeRangeRequest (Maybe WeekDay)

The day on which the time range ends.

instanceEventWindowTimeRangeRequest_startWeekDay :: Lens' InstanceEventWindowTimeRangeRequest (Maybe WeekDay)

The day on which the time range begins.

instanceEventWindowTimeRangeRequest_startHour :: Lens' InstanceEventWindowTimeRangeRequest (Maybe Natural)

The hour when the time range begins.