Amazonka.EC2.Types.InstanceStatusDetails

Description

Documentation

data InstanceStatusDetails

Describes the instance status.

See: newInstanceStatusDetails smart constructor.

Constructors

InstanceStatusDetails'

Fields

status :: Maybe StatusType

The status.

impairedSince :: Maybe ISO8601

The time when a status check failed. For an instance that was launched and impaired, this is the time when the instance was launched.

name :: Maybe StatusName

The type of instance status.

newInstanceStatusDetails :: InstanceStatusDetails

Create a value of InstanceStatusDetails with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:status:InstanceStatusDetails', instanceStatusDetails_status - The status.

$sel:impairedSince:InstanceStatusDetails', instanceStatusDetails_impairedSince - The time when a status check failed. For an instance that was launched and impaired, this is the time when the instance was launched.

$sel:name:InstanceStatusDetails', instanceStatusDetails_name - The type of instance status.

instanceStatusDetails_status :: Lens' InstanceStatusDetails (Maybe StatusType)

The status.

instanceStatusDetails_impairedSince :: Lens' InstanceStatusDetails (Maybe UTCTime)

The time when a status check failed. For an instance that was launched and impaired, this is the time when the instance was launched.

instanceStatusDetails_name :: Lens' InstanceStatusDetails (Maybe StatusName)

The type of instance status.