Amazonka.EC2.Types.InstanceTypeInfo

Description

Documentation

data InstanceTypeInfo

Describes the instance type.

See: newInstanceTypeInfo smart constructor.

Constructors

InstanceTypeInfo'

Fields

hypervisor :: Maybe InstanceTypeHypervisor

The hypervisor for the instance type.

currentGeneration :: Maybe Bool

Indicates whether the instance type is current generation.

memoryInfo :: Maybe MemoryInfo

Describes the memory for the instance type.

placementGroupInfo :: Maybe PlacementGroupInfo

Describes the placement group settings for the instance type.

supportedBootModes :: Maybe [BootModeType]

The supported boot modes. For more information, see Boot modes in the Amazon EC2 User Guide.

supportedRootDeviceTypes :: Maybe [RootDeviceType]

The supported root device types.

supportedUsageClasses :: Maybe [UsageClassType]

Indicates whether the instance type is offered for spot or On-Demand.

instanceStorageSupported :: Maybe Bool

Indicates whether instance storage is supported.

fpgaInfo :: Maybe FpgaInfo

Describes the FPGA accelerator settings for the instance type.

burstablePerformanceSupported :: Maybe Bool

Indicates whether the instance type is a burstable performance instance type.

instanceType :: Maybe InstanceType

The instance type. For more information, see Instance types in the Amazon EC2 User Guide.

gpuInfo :: Maybe GpuInfo

Describes the GPU accelerator settings for the instance type.

supportedVirtualizationTypes :: Maybe [VirtualizationType]

The supported virtualization types.

ebsInfo :: Maybe EbsInfo

Describes the Amazon EBS settings for the instance type.

autoRecoverySupported :: Maybe Bool

Indicates whether auto recovery is supported.

inferenceAcceleratorInfo :: Maybe InferenceAcceleratorInfo

Describes the Inference accelerator settings for the instance type.

bareMetal :: Maybe Bool

Indicates whether the instance is a bare metal instance type.

networkInfo :: Maybe NetworkInfo

Describes the network settings for the instance type.

processorInfo :: Maybe ProcessorInfo

Describes the processor.

freeTierEligible :: Maybe Bool

Indicates whether the instance type is eligible for the free tier.

vCpuInfo :: Maybe VCpuInfo

Describes the vCPU configurations for the instance type.

instanceStorageInfo :: Maybe InstanceStorageInfo

Describes the instance storage for the instance type.

dedicatedHostsSupported :: Maybe Bool

Indicates whether Dedicated Hosts are supported on the instance type.

hibernationSupported :: Maybe Bool

Indicates whether On-Demand hibernation is supported.

newInstanceTypeInfo :: InstanceTypeInfo

Create a value of InstanceTypeInfo with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:hypervisor:InstanceTypeInfo', instanceTypeInfo_hypervisor - The hypervisor for the instance type.

$sel:currentGeneration:InstanceTypeInfo', instanceTypeInfo_currentGeneration - Indicates whether the instance type is current generation.

$sel:memoryInfo:InstanceTypeInfo', instanceTypeInfo_memoryInfo - Describes the memory for the instance type.

$sel:placementGroupInfo:InstanceTypeInfo', instanceTypeInfo_placementGroupInfo - Describes the placement group settings for the instance type.

$sel:supportedBootModes:InstanceTypeInfo', instanceTypeInfo_supportedBootModes - The supported boot modes. For more information, see Boot modes in the Amazon EC2 User Guide.

$sel:supportedRootDeviceTypes:InstanceTypeInfo', instanceTypeInfo_supportedRootDeviceTypes - The supported root device types.

$sel:supportedUsageClasses:InstanceTypeInfo', instanceTypeInfo_supportedUsageClasses - Indicates whether the instance type is offered for spot or On-Demand.

$sel:instanceStorageSupported:InstanceTypeInfo', instanceTypeInfo_instanceStorageSupported - Indicates whether instance storage is supported.

$sel:fpgaInfo:InstanceTypeInfo', instanceTypeInfo_fpgaInfo - Describes the FPGA accelerator settings for the instance type.

$sel:burstablePerformanceSupported:InstanceTypeInfo', instanceTypeInfo_burstablePerformanceSupported - Indicates whether the instance type is a burstable performance instance type.

$sel:instanceType:InstanceTypeInfo', instanceTypeInfo_instanceType - The instance type. For more information, see Instance types in the Amazon EC2 User Guide.

$sel:gpuInfo:InstanceTypeInfo', instanceTypeInfo_gpuInfo - Describes the GPU accelerator settings for the instance type.

$sel:supportedVirtualizationTypes:InstanceTypeInfo', instanceTypeInfo_supportedVirtualizationTypes - The supported virtualization types.

$sel:ebsInfo:InstanceTypeInfo', instanceTypeInfo_ebsInfo - Describes the Amazon EBS settings for the instance type.

$sel:autoRecoverySupported:InstanceTypeInfo', instanceTypeInfo_autoRecoverySupported - Indicates whether auto recovery is supported.

$sel:inferenceAcceleratorInfo:InstanceTypeInfo', instanceTypeInfo_inferenceAcceleratorInfo - Describes the Inference accelerator settings for the instance type.

$sel:bareMetal:InstanceTypeInfo', instanceTypeInfo_bareMetal - Indicates whether the instance is a bare metal instance type.

$sel:networkInfo:InstanceTypeInfo', instanceTypeInfo_networkInfo - Describes the network settings for the instance type.

$sel:processorInfo:InstanceTypeInfo', instanceTypeInfo_processorInfo - Describes the processor.

$sel:freeTierEligible:InstanceTypeInfo', instanceTypeInfo_freeTierEligible - Indicates whether the instance type is eligible for the free tier.

$sel:vCpuInfo:InstanceTypeInfo', instanceTypeInfo_vCpuInfo - Describes the vCPU configurations for the instance type.

$sel:instanceStorageInfo:InstanceTypeInfo', instanceTypeInfo_instanceStorageInfo - Describes the instance storage for the instance type.

$sel:dedicatedHostsSupported:InstanceTypeInfo', instanceTypeInfo_dedicatedHostsSupported - Indicates whether Dedicated Hosts are supported on the instance type.

$sel:hibernationSupported:InstanceTypeInfo', instanceTypeInfo_hibernationSupported - Indicates whether On-Demand hibernation is supported.

instanceTypeInfo_hypervisor :: Lens' InstanceTypeInfo (Maybe InstanceTypeHypervisor)

The hypervisor for the instance type.

instanceTypeInfo_currentGeneration :: Lens' InstanceTypeInfo (Maybe Bool)

Indicates whether the instance type is current generation.

instanceTypeInfo_memoryInfo :: Lens' InstanceTypeInfo (Maybe MemoryInfo)

Describes the memory for the instance type.

instanceTypeInfo_placementGroupInfo :: Lens' InstanceTypeInfo (Maybe PlacementGroupInfo)

Describes the placement group settings for the instance type.

instanceTypeInfo_supportedBootModes :: Lens' InstanceTypeInfo (Maybe [BootModeType])

The supported boot modes. For more information, see Boot modes in the Amazon EC2 User Guide.

instanceTypeInfo_supportedRootDeviceTypes :: Lens' InstanceTypeInfo (Maybe [RootDeviceType])

The supported root device types.

instanceTypeInfo_supportedUsageClasses :: Lens' InstanceTypeInfo (Maybe [UsageClassType])

Indicates whether the instance type is offered for spot or On-Demand.

instanceTypeInfo_instanceStorageSupported :: Lens' InstanceTypeInfo (Maybe Bool)

Indicates whether instance storage is supported.

instanceTypeInfo_fpgaInfo :: Lens' InstanceTypeInfo (Maybe FpgaInfo)

Describes the FPGA accelerator settings for the instance type.

instanceTypeInfo_burstablePerformanceSupported :: Lens' InstanceTypeInfo (Maybe Bool)

Indicates whether the instance type is a burstable performance instance type.

instanceTypeInfo_instanceType :: Lens' InstanceTypeInfo (Maybe InstanceType)

The instance type. For more information, see Instance types in the Amazon EC2 User Guide.

instanceTypeInfo_gpuInfo :: Lens' InstanceTypeInfo (Maybe GpuInfo)

Describes the GPU accelerator settings for the instance type.

instanceTypeInfo_supportedVirtualizationTypes :: Lens' InstanceTypeInfo (Maybe [VirtualizationType])

The supported virtualization types.

instanceTypeInfo_ebsInfo :: Lens' InstanceTypeInfo (Maybe EbsInfo)

Describes the Amazon EBS settings for the instance type.

instanceTypeInfo_autoRecoverySupported :: Lens' InstanceTypeInfo (Maybe Bool)

Indicates whether auto recovery is supported.

instanceTypeInfo_inferenceAcceleratorInfo :: Lens' InstanceTypeInfo (Maybe InferenceAcceleratorInfo)

Describes the Inference accelerator settings for the instance type.

instanceTypeInfo_bareMetal :: Lens' InstanceTypeInfo (Maybe Bool)

Indicates whether the instance is a bare metal instance type.

instanceTypeInfo_networkInfo :: Lens' InstanceTypeInfo (Maybe NetworkInfo)

Describes the network settings for the instance type.

instanceTypeInfo_processorInfo :: Lens' InstanceTypeInfo (Maybe ProcessorInfo)

Describes the processor.

instanceTypeInfo_freeTierEligible :: Lens' InstanceTypeInfo (Maybe Bool)

Indicates whether the instance type is eligible for the free tier.

instanceTypeInfo_vCpuInfo :: Lens' InstanceTypeInfo (Maybe VCpuInfo)

Describes the vCPU configurations for the instance type.

instanceTypeInfo_instanceStorageInfo :: Lens' InstanceTypeInfo (Maybe InstanceStorageInfo)

Describes the instance storage for the instance type.

instanceTypeInfo_dedicatedHostsSupported :: Lens' InstanceTypeInfo (Maybe Bool)

Indicates whether Dedicated Hosts are supported on the instance type.

instanceTypeInfo_hibernationSupported :: Lens' InstanceTypeInfo (Maybe Bool)

Indicates whether On-Demand hibernation is supported.