Amazonka.EC2.Types.InternetGateway

Description

Documentation

data InternetGateway

Describes an internet gateway.

See: newInternetGateway smart constructor.

Constructors

InternetGateway'

Fields

attachments :: Maybe [InternetGatewayAttachment]

Any VPCs attached to the internet gateway.

ownerId :: Maybe Text

The ID of the Amazon Web Services account that owns the internet gateway.

tags :: Maybe [Tag]

Any tags assigned to the internet gateway.

internetGatewayId :: Text

The ID of the internet gateway.

newInternetGateway

Arguments

:: Text
$sel:internetGatewayId:InternetGateway'

-> InternetGateway

Create a value of InternetGateway with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:attachments:InternetGateway', internetGateway_attachments - Any VPCs attached to the internet gateway.

$sel:ownerId:InternetGateway', internetGateway_ownerId - The ID of the Amazon Web Services account that owns the internet gateway.

$sel:tags:InternetGateway', internetGateway_tags - Any tags assigned to the internet gateway.

$sel:internetGatewayId:InternetGateway', internetGateway_internetGatewayId - The ID of the internet gateway.

internetGateway_attachments :: Lens' InternetGateway (Maybe [InternetGatewayAttachment])

Any VPCs attached to the internet gateway.

internetGateway_ownerId :: Lens' InternetGateway (Maybe Text)

The ID of the Amazon Web Services account that owns the internet gateway.

internetGateway_tags :: Lens' InternetGateway (Maybe [Tag])

Any tags assigned to the internet gateway.

internetGateway_internetGatewayId :: Lens' InternetGateway Text

The ID of the internet gateway.