Amazonka.EC2.Types.KeyPairInfo

Description

Documentation

data KeyPairInfo

Describes a key pair.

See: newKeyPairInfo smart constructor.

Constructors

KeyPairInfo'

Fields

keyFingerprint :: Maybe Text

If you used CreateKeyPair to create the key pair:

For RSA key pairs, the key fingerprint is the SHA-1 digest of the DER encoded private key.

For ED25519 key pairs, the key fingerprint is the base64-encoded SHA-256 digest, which is the default for OpenSSH, starting with OpenSSH 6.8.

If you used ImportKeyPair to provide Amazon Web Services the public key:

For RSA key pairs, the key fingerprint is the MD5 public key fingerprint as specified in section 4 of RFC4716.

For ED25519 key pairs, the key fingerprint is the base64-encoded SHA-256 digest, which is the default for OpenSSH, starting with OpenSSH 6.8.

keyType :: Maybe KeyType

The type of key pair.

keyName :: Maybe Text

The name of the key pair.

keyPairId :: Maybe Text

The ID of the key pair.

tags :: Maybe [Tag]

Any tags applied to the key pair.

newKeyPairInfo :: KeyPairInfo

Create a value of KeyPairInfo with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:keyFingerprint:KeyPairInfo', keyPairInfo_keyFingerprint - If you used CreateKeyPair to create the key pair:

For RSA key pairs, the key fingerprint is the SHA-1 digest of the DER encoded private key.

For ED25519 key pairs, the key fingerprint is the base64-encoded SHA-256 digest, which is the default for OpenSSH, starting with OpenSSH 6.8.

If you used ImportKeyPair to provide Amazon Web Services the public key:

For RSA key pairs, the key fingerprint is the MD5 public key fingerprint as specified in section 4 of RFC4716.

For ED25519 key pairs, the key fingerprint is the base64-encoded SHA-256 digest, which is the default for OpenSSH, starting with OpenSSH 6.8.

$sel:keyType:KeyPairInfo', keyPairInfo_keyType - The type of key pair.

$sel:keyName:KeyPairInfo', keyPairInfo_keyName - The name of the key pair.

$sel:keyPairId:KeyPairInfo', keyPairInfo_keyPairId - The ID of the key pair.

$sel:tags:KeyPairInfo', keyPairInfo_tags - Any tags applied to the key pair.

keyPairInfo_keyFingerprint :: Lens' KeyPairInfo (Maybe Text)

If you used CreateKeyPair to create the key pair:

For RSA key pairs, the key fingerprint is the SHA-1 digest of the DER encoded private key.

For ED25519 key pairs, the key fingerprint is the base64-encoded SHA-256 digest, which is the default for OpenSSH, starting with OpenSSH 6.8.

If you used ImportKeyPair to provide Amazon Web Services the public key:

For RSA key pairs, the key fingerprint is the MD5 public key fingerprint as specified in section 4 of RFC4716.

For ED25519 key pairs, the key fingerprint is the base64-encoded SHA-256 digest, which is the default for OpenSSH, starting with OpenSSH 6.8.

keyPairInfo_keyType :: Lens' KeyPairInfo (Maybe KeyType)

The type of key pair.

keyPairInfo_keyName :: Lens' KeyPairInfo (Maybe Text)

The name of the key pair.

keyPairInfo_keyPairId :: Lens' KeyPairInfo (Maybe Text)

The ID of the key pair.

keyPairInfo_tags :: Lens' KeyPairInfo (Maybe [Tag])

Any tags applied to the key pair.