Amazonka.EC2.Types.LaunchSpecification

Description

Documentation

data LaunchSpecification

Describes the launch specification for an instance.

See: newLaunchSpecification smart constructor.

Constructors

LaunchSpecification'

Fields

securityGroups :: Maybe [GroupIdentifier]

One or more security groups. When requesting instances in a VPC, you must specify the IDs of the security groups. When requesting instances in EC2-Classic, you can specify the names or the IDs of the security groups.

keyName :: Maybe Text

The name of the key pair.

networkInterfaces :: Maybe [InstanceNetworkInterfaceSpecification]

One or more network interfaces. If you specify a network interface, you must specify subnet IDs and security group IDs using the network interface.

ramdiskId :: Maybe Text

The ID of the RAM disk.

subnetId :: Maybe Text

The ID of the subnet in which to launch the instance.

kernelId :: Maybe Text

The ID of the kernel.

instanceType :: Maybe InstanceType

The instance type.

ebsOptimized :: Maybe Bool

Indicates whether the instance is optimized for EBS I/O. This optimization provides dedicated throughput to Amazon EBS and an optimized configuration stack to provide optimal EBS I/O performance. This optimization isn't available with all instance types. Additional usage charges apply when using an EBS Optimized instance.

Default: false

userData :: Maybe Text

The Base64-encoded user data for the instance.

monitoring :: Maybe RunInstancesMonitoringEnabled

iamInstanceProfile :: Maybe IamInstanceProfileSpecification

The IAM instance profile.

imageId :: Maybe Text

The ID of the AMI.

addressingType :: Maybe Text

Deprecated.

blockDeviceMappings :: Maybe [BlockDeviceMapping]

One or more block device mapping entries.

placement :: Maybe SpotPlacement

The placement information for the instance.

newLaunchSpecification :: LaunchSpecification

Create a value of LaunchSpecification with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:securityGroups:LaunchSpecification', launchSpecification_securityGroups - One or more security groups. When requesting instances in a VPC, you must specify the IDs of the security groups. When requesting instances in EC2-Classic, you can specify the names or the IDs of the security groups.

$sel:keyName:LaunchSpecification', launchSpecification_keyName - The name of the key pair.

$sel:networkInterfaces:LaunchSpecification', launchSpecification_networkInterfaces - One or more network interfaces. If you specify a network interface, you must specify subnet IDs and security group IDs using the network interface.

$sel:ramdiskId:LaunchSpecification', launchSpecification_ramdiskId - The ID of the RAM disk.

$sel:subnetId:LaunchSpecification', launchSpecification_subnetId - The ID of the subnet in which to launch the instance.

$sel:kernelId:LaunchSpecification', launchSpecification_kernelId - The ID of the kernel.

$sel:instanceType:LaunchSpecification', launchSpecification_instanceType - The instance type.

$sel:ebsOptimized:LaunchSpecification', launchSpecification_ebsOptimized - Indicates whether the instance is optimized for EBS I/O. This optimization provides dedicated throughput to Amazon EBS and an optimized configuration stack to provide optimal EBS I/O performance. This optimization isn't available with all instance types. Additional usage charges apply when using an EBS Optimized instance.

Default: false

$sel:userData:LaunchSpecification', launchSpecification_userData - The Base64-encoded user data for the instance.

$sel:monitoring:LaunchSpecification', launchSpecification_monitoring - Undocumented member.

$sel:iamInstanceProfile:LaunchSpecification', launchSpecification_iamInstanceProfile - The IAM instance profile.

$sel:imageId:LaunchSpecification', launchSpecification_imageId - The ID of the AMI.

$sel:addressingType:LaunchSpecification', launchSpecification_addressingType - Deprecated.

$sel:blockDeviceMappings:LaunchSpecification', launchSpecification_blockDeviceMappings - One or more block device mapping entries.

$sel:placement:LaunchSpecification', launchSpecification_placement - The placement information for the instance.

launchSpecification_securityGroups :: Lens' LaunchSpecification (Maybe [GroupIdentifier])

One or more security groups. When requesting instances in a VPC, you must specify the IDs of the security groups. When requesting instances in EC2-Classic, you can specify the names or the IDs of the security groups.

launchSpecification_keyName :: Lens' LaunchSpecification (Maybe Text)

The name of the key pair.

launchSpecification_networkInterfaces :: Lens' LaunchSpecification (Maybe [InstanceNetworkInterfaceSpecification])

One or more network interfaces. If you specify a network interface, you must specify subnet IDs and security group IDs using the network interface.

launchSpecification_ramdiskId :: Lens' LaunchSpecification (Maybe Text)

The ID of the RAM disk.

launchSpecification_subnetId :: Lens' LaunchSpecification (Maybe Text)

The ID of the subnet in which to launch the instance.

launchSpecification_kernelId :: Lens' LaunchSpecification (Maybe Text)

The ID of the kernel.

launchSpecification_instanceType :: Lens' LaunchSpecification (Maybe InstanceType)

The instance type.

launchSpecification_ebsOptimized :: Lens' LaunchSpecification (Maybe Bool)

Indicates whether the instance is optimized for EBS I/O. This optimization provides dedicated throughput to Amazon EBS and an optimized configuration stack to provide optimal EBS I/O performance. This optimization isn't available with all instance types. Additional usage charges apply when using an EBS Optimized instance.

Default: false

launchSpecification_userData :: Lens' LaunchSpecification (Maybe Text)

The Base64-encoded user data for the instance.

launchSpecification_monitoring :: Lens' LaunchSpecification (Maybe RunInstancesMonitoringEnabled)

Undocumented member.

launchSpecification_iamInstanceProfile :: Lens' LaunchSpecification (Maybe IamInstanceProfileSpecification)

The IAM instance profile.

launchSpecification_imageId :: Lens' LaunchSpecification (Maybe Text)

The ID of the AMI.

launchSpecification_addressingType :: Lens' LaunchSpecification (Maybe Text)

Deprecated.

launchSpecification_blockDeviceMappings :: Lens' LaunchSpecification (Maybe [BlockDeviceMapping])

One or more block device mapping entries.

launchSpecification_placement :: Lens' LaunchSpecification (Maybe SpotPlacement)

The placement information for the instance.