Amazonka.EC2.Types.LaunchTemplate

Description

Documentation

data LaunchTemplate

Describes a launch template.

See: newLaunchTemplate smart constructor.

Constructors

LaunchTemplate'

Fields

launchTemplateName :: Maybe Text

The name of the launch template.

latestVersionNumber :: Maybe Integer

The version number of the latest version of the launch template.

launchTemplateId :: Maybe Text

The ID of the launch template.

createdBy :: Maybe Text

The principal that created the launch template.

defaultVersionNumber :: Maybe Integer

The version number of the default version of the launch template.

createTime :: Maybe ISO8601

The time launch template was created.

tags :: Maybe [Tag]

The tags for the launch template.

newLaunchTemplate :: LaunchTemplate

Create a value of LaunchTemplate with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:launchTemplateName:LaunchTemplate', launchTemplate_launchTemplateName - The name of the launch template.

$sel:latestVersionNumber:LaunchTemplate', launchTemplate_latestVersionNumber - The version number of the latest version of the launch template.

$sel:launchTemplateId:LaunchTemplate', launchTemplate_launchTemplateId - The ID of the launch template.

$sel:createdBy:LaunchTemplate', launchTemplate_createdBy - The principal that created the launch template.

$sel:defaultVersionNumber:LaunchTemplate', launchTemplate_defaultVersionNumber - The version number of the default version of the launch template.

$sel:createTime:LaunchTemplate', launchTemplate_createTime - The time launch template was created.

$sel:tags:LaunchTemplate', launchTemplate_tags - The tags for the launch template.

launchTemplate_launchTemplateName :: Lens' LaunchTemplate (Maybe Text)

The name of the launch template.

launchTemplate_latestVersionNumber :: Lens' LaunchTemplate (Maybe Integer)

The version number of the latest version of the launch template.

launchTemplate_launchTemplateId :: Lens' LaunchTemplate (Maybe Text)

The ID of the launch template.

launchTemplate_createdBy :: Lens' LaunchTemplate (Maybe Text)

The principal that created the launch template.

launchTemplate_defaultVersionNumber :: Lens' LaunchTemplate (Maybe Integer)

The version number of the default version of the launch template.

launchTemplate_createTime :: Lens' LaunchTemplate (Maybe UTCTime)

The time launch template was created.

launchTemplate_tags :: Lens' LaunchTemplate (Maybe [Tag])

The tags for the launch template.