Amazonka.EC2.Types.LaunchTemplateEbsBlockDeviceRequest

Description

Documentation

data LaunchTemplateEbsBlockDeviceRequest

The parameters for a block device for an EBS volume.

See: newLaunchTemplateEbsBlockDeviceRequest smart constructor.

Constructors

LaunchTemplateEbsBlockDeviceRequest'

Fields

deleteOnTermination :: Maybe Bool

Indicates whether the EBS volume is deleted on instance termination.

throughput :: Maybe Int

The throughput to provision for a gp3 volume, with a maximum of 1,000 MiB/s.

Valid Range: Minimum value of 125. Maximum value of 1000.

volumeSize :: Maybe Int

The size of the volume, in GiBs. You must specify either a snapshot ID or a volume size. The following are the supported volumes sizes for each volume type:

gp2 and gp3: 1-16,384

io1 and io2: 4-16,384

st1 and sc1: 125-16,384

standard: 1-1,024

iops :: Maybe Int

The number of I/O operations per second (IOPS). For gp3, io1, and io2 volumes, this represents the number of IOPS that are provisioned for the volume. For gp2 volumes, this represents the baseline performance of the volume and the rate at which the volume accumulates I/O credits for bursting.

The following are the supported values for each volume type:

gp3: 3,000-16,000 IOPS

io1: 100-64,000 IOPS

io2: 100-64,000 IOPS

For io1 and io2 volumes, we guarantee 64,000 IOPS only for Instances built on the Nitro System. Other instance families guarantee performance up to 32,000 IOPS.

This parameter is supported for io1, io2, and gp3 volumes only. This parameter is not supported for gp2, st1, sc1, or standard volumes.

encrypted :: Maybe Bool

Indicates whether the EBS volume is encrypted. Encrypted volumes can only be attached to instances that support Amazon EBS encryption. If you are creating a volume from a snapshot, you can't specify an encryption value.

kmsKeyId :: Maybe Text

The ARN of the symmetric Key Management Service (KMS) CMK used for encryption.

volumeType :: Maybe VolumeType

The volume type. For more information, see Amazon EBS volume types in the Amazon Elastic Compute Cloud User Guide.

snapshotId :: Maybe Text

The ID of the snapshot.

newLaunchTemplateEbsBlockDeviceRequest :: LaunchTemplateEbsBlockDeviceRequest

Create a value of LaunchTemplateEbsBlockDeviceRequest with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:deleteOnTermination:LaunchTemplateEbsBlockDeviceRequest', launchTemplateEbsBlockDeviceRequest_deleteOnTermination - Indicates whether the EBS volume is deleted on instance termination.

$sel:throughput:LaunchTemplateEbsBlockDeviceRequest', launchTemplateEbsBlockDeviceRequest_throughput - The throughput to provision for a gp3 volume, with a maximum of 1,000 MiB/s.

Valid Range: Minimum value of 125. Maximum value of 1000.

$sel:volumeSize:LaunchTemplateEbsBlockDeviceRequest', launchTemplateEbsBlockDeviceRequest_volumeSize - The size of the volume, in GiBs. You must specify either a snapshot ID or a volume size. The following are the supported volumes sizes for each volume type:

gp2 and gp3: 1-16,384

io1 and io2: 4-16,384

st1 and sc1: 125-16,384

standard: 1-1,024

$sel:iops:LaunchTemplateEbsBlockDeviceRequest', launchTemplateEbsBlockDeviceRequest_iops - The number of I/O operations per second (IOPS). For gp3, io1, and io2 volumes, this represents the number of IOPS that are provisioned for the volume. For gp2 volumes, this represents the baseline performance of the volume and the rate at which the volume accumulates I/O credits for bursting.

The following are the supported values for each volume type:

gp3: 3,000-16,000 IOPS

io1: 100-64,000 IOPS

io2: 100-64,000 IOPS

For io1 and io2 volumes, we guarantee 64,000 IOPS only for Instances built on the Nitro System. Other instance families guarantee performance up to 32,000 IOPS.

This parameter is supported for io1, io2, and gp3 volumes only. This parameter is not supported for gp2, st1, sc1, or standard volumes.

$sel:encrypted:LaunchTemplateEbsBlockDeviceRequest', launchTemplateEbsBlockDeviceRequest_encrypted - Indicates whether the EBS volume is encrypted. Encrypted volumes can only be attached to instances that support Amazon EBS encryption. If you are creating a volume from a snapshot, you can't specify an encryption value.

$sel:kmsKeyId:LaunchTemplateEbsBlockDeviceRequest', launchTemplateEbsBlockDeviceRequest_kmsKeyId - The ARN of the symmetric Key Management Service (KMS) CMK used for encryption.

$sel:volumeType:LaunchTemplateEbsBlockDeviceRequest', launchTemplateEbsBlockDeviceRequest_volumeType - The volume type. For more information, see Amazon EBS volume types in the Amazon Elastic Compute Cloud User Guide.

$sel:snapshotId:LaunchTemplateEbsBlockDeviceRequest', launchTemplateEbsBlockDeviceRequest_snapshotId - The ID of the snapshot.

launchTemplateEbsBlockDeviceRequest_deleteOnTermination :: Lens' LaunchTemplateEbsBlockDeviceRequest (Maybe Bool)

Indicates whether the EBS volume is deleted on instance termination.

launchTemplateEbsBlockDeviceRequest_throughput :: Lens' LaunchTemplateEbsBlockDeviceRequest (Maybe Int)

The throughput to provision for a gp3 volume, with a maximum of 1,000 MiB/s.

Valid Range: Minimum value of 125. Maximum value of 1000.

launchTemplateEbsBlockDeviceRequest_volumeSize :: Lens' LaunchTemplateEbsBlockDeviceRequest (Maybe Int)

The size of the volume, in GiBs. You must specify either a snapshot ID or a volume size. The following are the supported volumes sizes for each volume type:

gp2 and gp3: 1-16,384

io1 and io2: 4-16,384

st1 and sc1: 125-16,384

standard: 1-1,024

launchTemplateEbsBlockDeviceRequest_iops :: Lens' LaunchTemplateEbsBlockDeviceRequest (Maybe Int)

The number of I/O operations per second (IOPS). For gp3, io1, and io2 volumes, this represents the number of IOPS that are provisioned for the volume. For gp2 volumes, this represents the baseline performance of the volume and the rate at which the volume accumulates I/O credits for bursting.

The following are the supported values for each volume type:

gp3: 3,000-16,000 IOPS

io1: 100-64,000 IOPS

io2: 100-64,000 IOPS

For io1 and io2 volumes, we guarantee 64,000 IOPS only for Instances built on the Nitro System. Other instance families guarantee performance up to 32,000 IOPS.

This parameter is supported for io1, io2, and gp3 volumes only. This parameter is not supported for gp2, st1, sc1, or standard volumes.

launchTemplateEbsBlockDeviceRequest_encrypted :: Lens' LaunchTemplateEbsBlockDeviceRequest (Maybe Bool)

Indicates whether the EBS volume is encrypted. Encrypted volumes can only be attached to instances that support Amazon EBS encryption. If you are creating a volume from a snapshot, you can't specify an encryption value.

launchTemplateEbsBlockDeviceRequest_kmsKeyId :: Lens' LaunchTemplateEbsBlockDeviceRequest (Maybe Text)

The ARN of the symmetric Key Management Service (KMS) CMK used for encryption.

launchTemplateEbsBlockDeviceRequest_volumeType :: Lens' LaunchTemplateEbsBlockDeviceRequest (Maybe VolumeType)

The volume type. For more information, see Amazon EBS volume types in the Amazon Elastic Compute Cloud User Guide.

launchTemplateEbsBlockDeviceRequest_snapshotId :: Lens' LaunchTemplateEbsBlockDeviceRequest (Maybe Text)

The ID of the snapshot.