Amazonka.EC2.Types.LaunchTemplateOverrides

Description

Documentation

data LaunchTemplateOverrides

Describes overrides for a launch template.

See: newLaunchTemplateOverrides smart constructor.

Constructors

LaunchTemplateOverrides'

Fields

priority :: Maybe Double

The priority for the launch template override. The highest priority is launched first.

If OnDemandAllocationStrategy is set to prioritized, Spot Fleet uses priority to determine which launch template override to use first in fulfilling On-Demand capacity.

If the Spot AllocationStrategy is set to capacityOptimizedPrioritized, Spot Fleet uses priority on a best-effort basis to determine which launch template override to use in fulfilling Spot capacity, but optimizes for capacity first.

Valid values are whole numbers starting at 0. The lower the number, the higher the priority. If no number is set, the launch template override has the lowest priority. You can set the same priority for different launch template overrides.

spotPrice :: Maybe Text

The maximum price per unit hour that you are willing to pay for a Spot Instance.

weightedCapacity :: Maybe Double

The number of units provided by the specified instance type.

subnetId :: Maybe Text

The ID of the subnet in which to launch the instances.

instanceType :: Maybe InstanceType

The instance type.

availabilityZone :: Maybe Text

The Availability Zone in which to launch the instances.

newLaunchTemplateOverrides :: LaunchTemplateOverrides

Create a value of LaunchTemplateOverrides with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:priority:LaunchTemplateOverrides', launchTemplateOverrides_priority - The priority for the launch template override. The highest priority is launched first.

If OnDemandAllocationStrategy is set to prioritized, Spot Fleet uses priority to determine which launch template override to use first in fulfilling On-Demand capacity.

If the Spot AllocationStrategy is set to capacityOptimizedPrioritized, Spot Fleet uses priority on a best-effort basis to determine which launch template override to use in fulfilling Spot capacity, but optimizes for capacity first.

Valid values are whole numbers starting at 0. The lower the number, the higher the priority. If no number is set, the launch template override has the lowest priority. You can set the same priority for different launch template overrides.

$sel:spotPrice:LaunchTemplateOverrides', launchTemplateOverrides_spotPrice - The maximum price per unit hour that you are willing to pay for a Spot Instance.

$sel:weightedCapacity:LaunchTemplateOverrides', launchTemplateOverrides_weightedCapacity - The number of units provided by the specified instance type.

$sel:subnetId:LaunchTemplateOverrides', launchTemplateOverrides_subnetId - The ID of the subnet in which to launch the instances.

$sel:instanceType:LaunchTemplateOverrides', launchTemplateOverrides_instanceType - The instance type.

$sel:availabilityZone:LaunchTemplateOverrides', launchTemplateOverrides_availabilityZone - The Availability Zone in which to launch the instances.

launchTemplateOverrides_priority :: Lens' LaunchTemplateOverrides (Maybe Double)

The priority for the launch template override. The highest priority is launched first.

If OnDemandAllocationStrategy is set to prioritized, Spot Fleet uses priority to determine which launch template override to use first in fulfilling On-Demand capacity.

If the Spot AllocationStrategy is set to capacityOptimizedPrioritized, Spot Fleet uses priority on a best-effort basis to determine which launch template override to use in fulfilling Spot capacity, but optimizes for capacity first.

Valid values are whole numbers starting at 0. The lower the number, the higher the priority. If no number is set, the launch template override has the lowest priority. You can set the same priority for different launch template overrides.

launchTemplateOverrides_spotPrice :: Lens' LaunchTemplateOverrides (Maybe Text)

The maximum price per unit hour that you are willing to pay for a Spot Instance.

launchTemplateOverrides_weightedCapacity :: Lens' LaunchTemplateOverrides (Maybe Double)

The number of units provided by the specified instance type.

launchTemplateOverrides_subnetId :: Lens' LaunchTemplateOverrides (Maybe Text)

The ID of the subnet in which to launch the instances.

launchTemplateOverrides_instanceType :: Lens' LaunchTemplateOverrides (Maybe InstanceType)

The instance type.

launchTemplateOverrides_availabilityZone :: Lens' LaunchTemplateOverrides (Maybe Text)

The Availability Zone in which to launch the instances.