Amazonka.EC2.Types.LaunchTemplatePlacement

Description

Documentation

data LaunchTemplatePlacement

Describes the placement of an instance.

See: newLaunchTemplatePlacement smart constructor.

Constructors

LaunchTemplatePlacement'

Fields

affinity :: Maybe Text

The affinity setting for the instance on the Dedicated Host.

hostId :: Maybe Text

The ID of the Dedicated Host for the instance.

partitionNumber :: Maybe Int

The number of the partition the instance should launch in. Valid only if the placement group strategy is set to partition.

spreadDomain :: Maybe Text

Reserved for future use.

availabilityZone :: Maybe Text

The Availability Zone of the instance.

tenancy :: Maybe Tenancy

The tenancy of the instance (if the instance is running in a VPC). An instance with a tenancy of dedicated runs on single-tenant hardware.

groupName :: Maybe Text

The name of the placement group for the instance.

hostResourceGroupArn :: Maybe Text

The ARN of the host resource group in which to launch the instances.

newLaunchTemplatePlacement :: LaunchTemplatePlacement

Create a value of LaunchTemplatePlacement with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:affinity:LaunchTemplatePlacement', launchTemplatePlacement_affinity - The affinity setting for the instance on the Dedicated Host.

$sel:hostId:LaunchTemplatePlacement', launchTemplatePlacement_hostId - The ID of the Dedicated Host for the instance.

$sel:partitionNumber:LaunchTemplatePlacement', launchTemplatePlacement_partitionNumber - The number of the partition the instance should launch in. Valid only if the placement group strategy is set to partition.

$sel:spreadDomain:LaunchTemplatePlacement', launchTemplatePlacement_spreadDomain - Reserved for future use.

$sel:availabilityZone:LaunchTemplatePlacement', launchTemplatePlacement_availabilityZone - The Availability Zone of the instance.

$sel:tenancy:LaunchTemplatePlacement', launchTemplatePlacement_tenancy - The tenancy of the instance (if the instance is running in a VPC). An instance with a tenancy of dedicated runs on single-tenant hardware.

$sel:groupName:LaunchTemplatePlacement', launchTemplatePlacement_groupName - The name of the placement group for the instance.

$sel:hostResourceGroupArn:LaunchTemplatePlacement', launchTemplatePlacement_hostResourceGroupArn - The ARN of the host resource group in which to launch the instances.

launchTemplatePlacement_affinity :: Lens' LaunchTemplatePlacement (Maybe Text)

The affinity setting for the instance on the Dedicated Host.

launchTemplatePlacement_hostId :: Lens' LaunchTemplatePlacement (Maybe Text)

The ID of the Dedicated Host for the instance.

launchTemplatePlacement_partitionNumber :: Lens' LaunchTemplatePlacement (Maybe Int)

The number of the partition the instance should launch in. Valid only if the placement group strategy is set to partition.

launchTemplatePlacement_spreadDomain :: Lens' LaunchTemplatePlacement (Maybe Text)

Reserved for future use.

launchTemplatePlacement_availabilityZone :: Lens' LaunchTemplatePlacement (Maybe Text)

The Availability Zone of the instance.

launchTemplatePlacement_tenancy :: Lens' LaunchTemplatePlacement (Maybe Tenancy)

The tenancy of the instance (if the instance is running in a VPC). An instance with a tenancy of dedicated runs on single-tenant hardware.

launchTemplatePlacement_groupName :: Lens' LaunchTemplatePlacement (Maybe Text)

The name of the placement group for the instance.

launchTemplatePlacement_hostResourceGroupArn :: Lens' LaunchTemplatePlacement (Maybe Text)

The ARN of the host resource group in which to launch the instances.