Amazonka.EC2.Types.LaunchTemplatePlacementRequest

Description

Documentation

data LaunchTemplatePlacementRequest

Describes the placement of an instance.

See: newLaunchTemplatePlacementRequest smart constructor.

Constructors

LaunchTemplatePlacementRequest'

Fields

affinity :: Maybe Text

The affinity setting for an instance on a Dedicated Host.

hostId :: Maybe Text

The ID of the Dedicated Host for the instance.

partitionNumber :: Maybe Int

The number of the partition the instance should launch in. Valid only if the placement group strategy is set to partition.

spreadDomain :: Maybe Text

Reserved for future use.

availabilityZone :: Maybe Text

The Availability Zone for the instance.

tenancy :: Maybe Tenancy

The tenancy of the instance (if the instance is running in a VPC). An instance with a tenancy of dedicated runs on single-tenant hardware.

groupName :: Maybe Text

The name of the placement group for the instance.

hostResourceGroupArn :: Maybe Text

The ARN of the host resource group in which to launch the instances. If you specify a host resource group ARN, omit the Tenancy parameter or set it to host.

newLaunchTemplatePlacementRequest :: LaunchTemplatePlacementRequest

Create a value of LaunchTemplatePlacementRequest with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:affinity:LaunchTemplatePlacementRequest', launchTemplatePlacementRequest_affinity - The affinity setting for an instance on a Dedicated Host.

$sel:hostId:LaunchTemplatePlacementRequest', launchTemplatePlacementRequest_hostId - The ID of the Dedicated Host for the instance.

$sel:partitionNumber:LaunchTemplatePlacementRequest', launchTemplatePlacementRequest_partitionNumber - The number of the partition the instance should launch in. Valid only if the placement group strategy is set to partition.

$sel:spreadDomain:LaunchTemplatePlacementRequest', launchTemplatePlacementRequest_spreadDomain - Reserved for future use.

$sel:availabilityZone:LaunchTemplatePlacementRequest', launchTemplatePlacementRequest_availabilityZone - The Availability Zone for the instance.

$sel:tenancy:LaunchTemplatePlacementRequest', launchTemplatePlacementRequest_tenancy - The tenancy of the instance (if the instance is running in a VPC). An instance with a tenancy of dedicated runs on single-tenant hardware.

$sel:groupName:LaunchTemplatePlacementRequest', launchTemplatePlacementRequest_groupName - The name of the placement group for the instance.

$sel:hostResourceGroupArn:LaunchTemplatePlacementRequest', launchTemplatePlacementRequest_hostResourceGroupArn - The ARN of the host resource group in which to launch the instances. If you specify a host resource group ARN, omit the Tenancy parameter or set it to host.

launchTemplatePlacementRequest_affinity :: Lens' LaunchTemplatePlacementRequest (Maybe Text)

The affinity setting for an instance on a Dedicated Host.

launchTemplatePlacementRequest_hostId :: Lens' LaunchTemplatePlacementRequest (Maybe Text)

The ID of the Dedicated Host for the instance.

launchTemplatePlacementRequest_partitionNumber :: Lens' LaunchTemplatePlacementRequest (Maybe Int)

The number of the partition the instance should launch in. Valid only if the placement group strategy is set to partition.

launchTemplatePlacementRequest_spreadDomain :: Lens' LaunchTemplatePlacementRequest (Maybe Text)

Reserved for future use.

launchTemplatePlacementRequest_availabilityZone :: Lens' LaunchTemplatePlacementRequest (Maybe Text)

The Availability Zone for the instance.

launchTemplatePlacementRequest_tenancy :: Lens' LaunchTemplatePlacementRequest (Maybe Tenancy)

The tenancy of the instance (if the instance is running in a VPC). An instance with a tenancy of dedicated runs on single-tenant hardware.

launchTemplatePlacementRequest_groupName :: Lens' LaunchTemplatePlacementRequest (Maybe Text)

The name of the placement group for the instance.

launchTemplatePlacementRequest_hostResourceGroupArn :: Lens' LaunchTemplatePlacementRequest (Maybe Text)

The ARN of the host resource group in which to launch the instances. If you specify a host resource group ARN, omit the Tenancy parameter or set it to host.