Amazonka.EC2.Types.LocalGatewayRouteTableVirtualInterfaceGroupAssociation

Description

Documentation

data LocalGatewayRouteTableVirtualInterfaceGroupAssociation

Describes an association between a local gateway route table and a virtual interface group.

See: newLocalGatewayRouteTableVirtualInterfaceGroupAssociation smart constructor.

Constructors

LocalGatewayRouteTableVirtualInterfaceGroupAssociation'

Fields

state :: Maybe Text

The state of the association.

localGatewayRouteTableArn :: Maybe Text

The Amazon Resource Name (ARN) of the local gateway route table for the virtual interface group.

localGatewayId :: Maybe Text

The ID of the local gateway.

ownerId :: Maybe Text

The ID of the Amazon Web Services account that owns the local gateway virtual interface group association.

localGatewayRouteTableId :: Maybe Text

The ID of the local gateway route table.

localGatewayRouteTableVirtualInterfaceGroupAssociationId :: Maybe Text

The ID of the association.

localGatewayVirtualInterfaceGroupId :: Maybe Text

The ID of the virtual interface group.

tags :: Maybe [Tag]

The tags assigned to the association.

newLocalGatewayRouteTableVirtualInterfaceGroupAssociation :: LocalGatewayRouteTableVirtualInterfaceGroupAssociation

Create a value of LocalGatewayRouteTableVirtualInterfaceGroupAssociation with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:state:LocalGatewayRouteTableVirtualInterfaceGroupAssociation', localGatewayRouteTableVirtualInterfaceGroupAssociation_state - The state of the association.

$sel:localGatewayRouteTableArn:LocalGatewayRouteTableVirtualInterfaceGroupAssociation', localGatewayRouteTableVirtualInterfaceGroupAssociation_localGatewayRouteTableArn - The Amazon Resource Name (ARN) of the local gateway route table for the virtual interface group.

$sel:localGatewayId:LocalGatewayRouteTableVirtualInterfaceGroupAssociation', localGatewayRouteTableVirtualInterfaceGroupAssociation_localGatewayId - The ID of the local gateway.

$sel:ownerId:LocalGatewayRouteTableVirtualInterfaceGroupAssociation', localGatewayRouteTableVirtualInterfaceGroupAssociation_ownerId - The ID of the Amazon Web Services account that owns the local gateway virtual interface group association.

$sel:localGatewayRouteTableId:LocalGatewayRouteTableVirtualInterfaceGroupAssociation', localGatewayRouteTableVirtualInterfaceGroupAssociation_localGatewayRouteTableId - The ID of the local gateway route table.

$sel:localGatewayRouteTableVirtualInterfaceGroupAssociationId:LocalGatewayRouteTableVirtualInterfaceGroupAssociation', localGatewayRouteTableVirtualInterfaceGroupAssociation_localGatewayRouteTableVirtualInterfaceGroupAssociationId - The ID of the association.

$sel:localGatewayVirtualInterfaceGroupId:LocalGatewayRouteTableVirtualInterfaceGroupAssociation', localGatewayRouteTableVirtualInterfaceGroupAssociation_localGatewayVirtualInterfaceGroupId - The ID of the virtual interface group.

$sel:tags:LocalGatewayRouteTableVirtualInterfaceGroupAssociation', localGatewayRouteTableVirtualInterfaceGroupAssociation_tags - The tags assigned to the association.

localGatewayRouteTableVirtualInterfaceGroupAssociation_state :: Lens' LocalGatewayRouteTableVirtualInterfaceGroupAssociation (Maybe Text)

The state of the association.

localGatewayRouteTableVirtualInterfaceGroupAssociation_localGatewayRouteTableArn :: Lens' LocalGatewayRouteTableVirtualInterfaceGroupAssociation (Maybe Text)

The Amazon Resource Name (ARN) of the local gateway route table for the virtual interface group.

localGatewayRouteTableVirtualInterfaceGroupAssociation_localGatewayId :: Lens' LocalGatewayRouteTableVirtualInterfaceGroupAssociation (Maybe Text)

The ID of the local gateway.

localGatewayRouteTableVirtualInterfaceGroupAssociation_ownerId :: Lens' LocalGatewayRouteTableVirtualInterfaceGroupAssociation (Maybe Text)

The ID of the Amazon Web Services account that owns the local gateway virtual interface group association.

localGatewayRouteTableVirtualInterfaceGroupAssociation_localGatewayRouteTableId :: Lens' LocalGatewayRouteTableVirtualInterfaceGroupAssociation (Maybe Text)

The ID of the local gateway route table.

localGatewayRouteTableVirtualInterfaceGroupAssociation_localGatewayRouteTableVirtualInterfaceGroupAssociationId :: Lens' LocalGatewayRouteTableVirtualInterfaceGroupAssociation (Maybe Text)

The ID of the association.

localGatewayRouteTableVirtualInterfaceGroupAssociation_localGatewayVirtualInterfaceGroupId :: Lens' LocalGatewayRouteTableVirtualInterfaceGroupAssociation (Maybe Text)

The ID of the virtual interface group.

localGatewayRouteTableVirtualInterfaceGroupAssociation_tags :: Lens' LocalGatewayRouteTableVirtualInterfaceGroupAssociation (Maybe [Tag])

The tags assigned to the association.