Amazonka.EC2.Types.LocalGatewayVirtualInterface

Description

Documentation

data LocalGatewayVirtualInterface

Describes a local gateway virtual interface.

See: newLocalGatewayVirtualInterface smart constructor.

Constructors

LocalGatewayVirtualInterface'

Fields

localGatewayVirtualInterfaceId :: Maybe Text

The ID of the virtual interface.

localBgpAsn :: Maybe Int

The Border Gateway Protocol (BGP) Autonomous System Number (ASN) of the local gateway.

vlan :: Maybe Int

The ID of the VLAN.

localGatewayId :: Maybe Text

The ID of the local gateway.

localAddress :: Maybe Text

The local address.

peerBgpAsn :: Maybe Int

The peer BGP ASN.

ownerId :: Maybe Text

The ID of the Amazon Web Services account that owns the local gateway virtual interface.

peerAddress :: Maybe Text

The peer address.

tags :: Maybe [Tag]

The tags assigned to the virtual interface.

newLocalGatewayVirtualInterface :: LocalGatewayVirtualInterface

Create a value of LocalGatewayVirtualInterface with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:localGatewayVirtualInterfaceId:LocalGatewayVirtualInterface', localGatewayVirtualInterface_localGatewayVirtualInterfaceId - The ID of the virtual interface.

$sel:localBgpAsn:LocalGatewayVirtualInterface', localGatewayVirtualInterface_localBgpAsn - The Border Gateway Protocol (BGP) Autonomous System Number (ASN) of the local gateway.

$sel:vlan:LocalGatewayVirtualInterface', localGatewayVirtualInterface_vlan - The ID of the VLAN.

$sel:localGatewayId:LocalGatewayVirtualInterface', localGatewayVirtualInterface_localGatewayId - The ID of the local gateway.

$sel:localAddress:LocalGatewayVirtualInterface', localGatewayVirtualInterface_localAddress - The local address.

$sel:peerBgpAsn:LocalGatewayVirtualInterface', localGatewayVirtualInterface_peerBgpAsn - The peer BGP ASN.

$sel:ownerId:LocalGatewayVirtualInterface', localGatewayVirtualInterface_ownerId - The ID of the Amazon Web Services account that owns the local gateway virtual interface.

$sel:peerAddress:LocalGatewayVirtualInterface', localGatewayVirtualInterface_peerAddress - The peer address.

$sel:tags:LocalGatewayVirtualInterface', localGatewayVirtualInterface_tags - The tags assigned to the virtual interface.

localGatewayVirtualInterface_localGatewayVirtualInterfaceId :: Lens' LocalGatewayVirtualInterface (Maybe Text)

The ID of the virtual interface.

localGatewayVirtualInterface_localBgpAsn :: Lens' LocalGatewayVirtualInterface (Maybe Int)

The Border Gateway Protocol (BGP) Autonomous System Number (ASN) of the local gateway.

localGatewayVirtualInterface_vlan :: Lens' LocalGatewayVirtualInterface (Maybe Int)

The ID of the VLAN.

localGatewayVirtualInterface_localGatewayId :: Lens' LocalGatewayVirtualInterface (Maybe Text)

The ID of the local gateway.

localGatewayVirtualInterface_localAddress :: Lens' LocalGatewayVirtualInterface (Maybe Text)

The local address.

localGatewayVirtualInterface_peerBgpAsn :: Lens' LocalGatewayVirtualInterface (Maybe Int)

The peer BGP ASN.

localGatewayVirtualInterface_ownerId :: Lens' LocalGatewayVirtualInterface (Maybe Text)

The ID of the Amazon Web Services account that owns the local gateway virtual interface.

localGatewayVirtualInterface_peerAddress :: Lens' LocalGatewayVirtualInterface (Maybe Text)

The peer address.

localGatewayVirtualInterface_tags :: Lens' LocalGatewayVirtualInterface (Maybe [Tag])

The tags assigned to the virtual interface.