Amazonka.EC2.Types.ModifyTransitGatewayOptions

Description

Documentation

data ModifyTransitGatewayOptions

The transit gateway options.

See: newModifyTransitGatewayOptions smart constructor.

Constructors

ModifyTransitGatewayOptions'

Fields

vpnEcmpSupport :: Maybe VpnEcmpSupportValue

Enable or disable Equal Cost Multipath Protocol support.

autoAcceptSharedAttachments :: Maybe AutoAcceptSharedAttachmentsValue

Enable or disable automatic acceptance of attachment requests.

propagationDefaultRouteTableId :: Maybe Text

The ID of the default propagation route table.

removeTransitGatewayCidrBlocks :: Maybe [Text]

Removes CIDR blocks for the transit gateway.

defaultRouteTableAssociation :: Maybe DefaultRouteTableAssociationValue

Enable or disable automatic association with the default association route table.

associationDefaultRouteTableId :: Maybe Text

The ID of the default association route table.

defaultRouteTablePropagation :: Maybe DefaultRouteTablePropagationValue

Enable or disable automatic propagation of routes to the default propagation route table.

addTransitGatewayCidrBlocks :: Maybe [Text]

Adds IPv4 or IPv6 CIDR blocks for the transit gateway. Must be a size /24 CIDR block or larger for IPv4, or a size /64 CIDR block or larger for IPv6.

dnsSupport :: Maybe DnsSupportValue

Enable or disable DNS support.

newModifyTransitGatewayOptions :: ModifyTransitGatewayOptions

Create a value of ModifyTransitGatewayOptions with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:vpnEcmpSupport:ModifyTransitGatewayOptions', modifyTransitGatewayOptions_vpnEcmpSupport - Enable or disable Equal Cost Multipath Protocol support.

$sel:autoAcceptSharedAttachments:ModifyTransitGatewayOptions', modifyTransitGatewayOptions_autoAcceptSharedAttachments - Enable or disable automatic acceptance of attachment requests.

$sel:propagationDefaultRouteTableId:ModifyTransitGatewayOptions', modifyTransitGatewayOptions_propagationDefaultRouteTableId - The ID of the default propagation route table.

$sel:removeTransitGatewayCidrBlocks:ModifyTransitGatewayOptions', modifyTransitGatewayOptions_removeTransitGatewayCidrBlocks - Removes CIDR blocks for the transit gateway.

$sel:defaultRouteTableAssociation:ModifyTransitGatewayOptions', modifyTransitGatewayOptions_defaultRouteTableAssociation - Enable or disable automatic association with the default association route table.

$sel:associationDefaultRouteTableId:ModifyTransitGatewayOptions', modifyTransitGatewayOptions_associationDefaultRouteTableId - The ID of the default association route table.

$sel:defaultRouteTablePropagation:ModifyTransitGatewayOptions', modifyTransitGatewayOptions_defaultRouteTablePropagation - Enable or disable automatic propagation of routes to the default propagation route table.

$sel:addTransitGatewayCidrBlocks:ModifyTransitGatewayOptions', modifyTransitGatewayOptions_addTransitGatewayCidrBlocks - Adds IPv4 or IPv6 CIDR blocks for the transit gateway. Must be a size /24 CIDR block or larger for IPv4, or a size /64 CIDR block or larger for IPv6.

$sel:dnsSupport:ModifyTransitGatewayOptions', modifyTransitGatewayOptions_dnsSupport - Enable or disable DNS support.

modifyTransitGatewayOptions_vpnEcmpSupport :: Lens' ModifyTransitGatewayOptions (Maybe VpnEcmpSupportValue)

Enable or disable Equal Cost Multipath Protocol support.

modifyTransitGatewayOptions_autoAcceptSharedAttachments :: Lens' ModifyTransitGatewayOptions (Maybe AutoAcceptSharedAttachmentsValue)

Enable or disable automatic acceptance of attachment requests.

modifyTransitGatewayOptions_propagationDefaultRouteTableId :: Lens' ModifyTransitGatewayOptions (Maybe Text)

The ID of the default propagation route table.

modifyTransitGatewayOptions_removeTransitGatewayCidrBlocks :: Lens' ModifyTransitGatewayOptions (Maybe [Text])

Removes CIDR blocks for the transit gateway.

modifyTransitGatewayOptions_defaultRouteTableAssociation :: Lens' ModifyTransitGatewayOptions (Maybe DefaultRouteTableAssociationValue)

Enable or disable automatic association with the default association route table.

modifyTransitGatewayOptions_associationDefaultRouteTableId :: Lens' ModifyTransitGatewayOptions (Maybe Text)

The ID of the default association route table.

modifyTransitGatewayOptions_defaultRouteTablePropagation :: Lens' ModifyTransitGatewayOptions (Maybe DefaultRouteTablePropagationValue)

Enable or disable automatic propagation of routes to the default propagation route table.

modifyTransitGatewayOptions_addTransitGatewayCidrBlocks :: Lens' ModifyTransitGatewayOptions (Maybe [Text])

Adds IPv4 or IPv6 CIDR blocks for the transit gateway. Must be a size /24 CIDR block or larger for IPv4, or a size /64 CIDR block or larger for IPv6.

modifyTransitGatewayOptions_dnsSupport :: Lens' ModifyTransitGatewayOptions (Maybe DnsSupportValue)

Enable or disable DNS support.