Amazonka.EC2.Types.NetworkInfo

Description

Documentation

data NetworkInfo

Describes the networking features of the instance type.

See: newNetworkInfo smart constructor.

Constructors

NetworkInfo'

Fields

efaSupported :: Maybe Bool

Indicates whether Elastic Fabric Adapter (EFA) is supported.

ipv6Supported :: Maybe Bool

Indicates whether IPv6 is supported.

enaSupport :: Maybe EnaSupport

Indicates whether Elastic Network Adapter (ENA) is supported.

maximumNetworkInterfaces :: Maybe Int

The maximum number of network interfaces for the instance type.

ipv6AddressesPerInterface :: Maybe Int

The maximum number of IPv6 addresses per network interface.

networkPerformance :: Maybe Text

The network performance.

maximumNetworkCards :: Maybe Int

The maximum number of physical network cards that can be allocated to the instance.

networkCards :: Maybe [NetworkCardInfo]

Describes the network cards for the instance type.

efaInfo :: Maybe EfaInfo

Describes the Elastic Fabric Adapters for the instance type.

defaultNetworkCardIndex :: Maybe Int

The index of the default network card, starting at 0.

ipv4AddressesPerInterface :: Maybe Int

The maximum number of IPv4 addresses per network interface.

encryptionInTransitSupported :: Maybe Bool

Indicates whether the instance type automatically encrypts in-transit traffic between instances.

newNetworkInfo :: NetworkInfo

Create a value of NetworkInfo with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:efaSupported:NetworkInfo', networkInfo_efaSupported - Indicates whether Elastic Fabric Adapter (EFA) is supported.

$sel:ipv6Supported:NetworkInfo', networkInfo_ipv6Supported - Indicates whether IPv6 is supported.

$sel:enaSupport:NetworkInfo', networkInfo_enaSupport - Indicates whether Elastic Network Adapter (ENA) is supported.

$sel:maximumNetworkInterfaces:NetworkInfo', networkInfo_maximumNetworkInterfaces - The maximum number of network interfaces for the instance type.

$sel:ipv6AddressesPerInterface:NetworkInfo', networkInfo_ipv6AddressesPerInterface - The maximum number of IPv6 addresses per network interface.

$sel:networkPerformance:NetworkInfo', networkInfo_networkPerformance - The network performance.

$sel:maximumNetworkCards:NetworkInfo', networkInfo_maximumNetworkCards - The maximum number of physical network cards that can be allocated to the instance.

$sel:networkCards:NetworkInfo', networkInfo_networkCards - Describes the network cards for the instance type.

$sel:efaInfo:NetworkInfo', networkInfo_efaInfo - Describes the Elastic Fabric Adapters for the instance type.

$sel:defaultNetworkCardIndex:NetworkInfo', networkInfo_defaultNetworkCardIndex - The index of the default network card, starting at 0.

$sel:ipv4AddressesPerInterface:NetworkInfo', networkInfo_ipv4AddressesPerInterface - The maximum number of IPv4 addresses per network interface.

$sel:encryptionInTransitSupported:NetworkInfo', networkInfo_encryptionInTransitSupported - Indicates whether the instance type automatically encrypts in-transit traffic between instances.

networkInfo_efaSupported :: Lens' NetworkInfo (Maybe Bool)

Indicates whether Elastic Fabric Adapter (EFA) is supported.

networkInfo_ipv6Supported :: Lens' NetworkInfo (Maybe Bool)

Indicates whether IPv6 is supported.

networkInfo_enaSupport :: Lens' NetworkInfo (Maybe EnaSupport)

Indicates whether Elastic Network Adapter (ENA) is supported.

networkInfo_maximumNetworkInterfaces :: Lens' NetworkInfo (Maybe Int)

The maximum number of network interfaces for the instance type.

networkInfo_ipv6AddressesPerInterface :: Lens' NetworkInfo (Maybe Int)

The maximum number of IPv6 addresses per network interface.

networkInfo_networkPerformance :: Lens' NetworkInfo (Maybe Text)

The network performance.

networkInfo_maximumNetworkCards :: Lens' NetworkInfo (Maybe Int)

The maximum number of physical network cards that can be allocated to the instance.

networkInfo_networkCards :: Lens' NetworkInfo (Maybe [NetworkCardInfo])

Describes the network cards for the instance type.

networkInfo_efaInfo :: Lens' NetworkInfo (Maybe EfaInfo)

Describes the Elastic Fabric Adapters for the instance type.

networkInfo_defaultNetworkCardIndex :: Lens' NetworkInfo (Maybe Int)

The index of the default network card, starting at 0.

networkInfo_ipv4AddressesPerInterface :: Lens' NetworkInfo (Maybe Int)

The maximum number of IPv4 addresses per network interface.

networkInfo_encryptionInTransitSupported :: Lens' NetworkInfo (Maybe Bool)

Indicates whether the instance type automatically encrypts in-transit traffic between instances.