Amazonka.EC2.Types.NetworkInterface

Description

Documentation

data NetworkInterface

Describes a network interface.

See: newNetworkInterface smart constructor.

Constructors

NetworkInterface'

Fields

groups :: Maybe [GroupIdentifier]

Any security groups for the network interface.

status :: Maybe NetworkInterfaceStatus

The status of the network interface.

privateIpAddresses :: Maybe [NetworkInterfacePrivateIpAddress]

The private IPv4 addresses associated with the network interface.

ipv4Prefixes :: Maybe [Ipv4PrefixSpecification]

The IPv4 prefixes that are assigned to the network interface.

sourceDestCheck :: Maybe Bool

Indicates whether source/destination checking is enabled.

interfaceType :: Maybe NetworkInterfaceType

The type of network interface.

vpcId :: Maybe Text

The ID of the VPC.

tagSet :: Maybe [Tag]

Any tags assigned to the network interface.

requesterManaged :: Maybe Bool

Indicates whether the network interface is being managed by Amazon Web Services.

outpostArn :: Maybe Text

The Amazon Resource Name (ARN) of the Outpost.

networkInterfaceId :: Maybe Text

The ID of the network interface.

subnetId :: Maybe Text

The ID of the subnet.

macAddress :: Maybe Text

The MAC address.

attachment :: Maybe NetworkInterfaceAttachment

The network interface attachment.

ownerId :: Maybe Text

The Amazon Web Services account ID of the owner of the network interface.

ipv6Prefixes :: Maybe [Ipv6PrefixSpecification]

The IPv6 prefixes that are assigned to the network interface.

availabilityZone :: Maybe Text

The Availability Zone.

privateIpAddress :: Maybe Text

The IPv4 address of the network interface within the subnet.

privateDnsName :: Maybe Text

The private DNS name.

requesterId :: Maybe Text

The alias or Amazon Web Services account ID of the principal or service that created the network interface.

description :: Maybe Text

A description.

association :: Maybe NetworkInterfaceAssociation

The association information for an Elastic IP address (IPv4) associated with the network interface.

ipv6Addresses :: Maybe [NetworkInterfaceIpv6Address]

The IPv6 addresses associated with the network interface.

newNetworkInterface :: NetworkInterface

Create a value of NetworkInterface with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:groups:NetworkInterface', networkInterface_groups - Any security groups for the network interface.

$sel:status:NetworkInterface', networkInterface_status - The status of the network interface.

$sel:privateIpAddresses:NetworkInterface', networkInterface_privateIpAddresses - The private IPv4 addresses associated with the network interface.

$sel:ipv4Prefixes:NetworkInterface', networkInterface_ipv4Prefixes - The IPv4 prefixes that are assigned to the network interface.

$sel:sourceDestCheck:NetworkInterface', networkInterface_sourceDestCheck - Indicates whether source/destination checking is enabled.

$sel:interfaceType:NetworkInterface', networkInterface_interfaceType - The type of network interface.

$sel:vpcId:NetworkInterface', networkInterface_vpcId - The ID of the VPC.

$sel:tagSet:NetworkInterface', networkInterface_tagSet - Any tags assigned to the network interface.

$sel:requesterManaged:NetworkInterface', networkInterface_requesterManaged - Indicates whether the network interface is being managed by Amazon Web Services.

$sel:outpostArn:NetworkInterface', networkInterface_outpostArn - The Amazon Resource Name (ARN) of the Outpost.

$sel:networkInterfaceId:NetworkInterface', networkInterface_networkInterfaceId - The ID of the network interface.

$sel:subnetId:NetworkInterface', networkInterface_subnetId - The ID of the subnet.

$sel:macAddress:NetworkInterface', networkInterface_macAddress - The MAC address.

$sel:attachment:NetworkInterface', networkInterface_attachment - The network interface attachment.

$sel:ownerId:NetworkInterface', networkInterface_ownerId - The Amazon Web Services account ID of the owner of the network interface.

$sel:ipv6Prefixes:NetworkInterface', networkInterface_ipv6Prefixes - The IPv6 prefixes that are assigned to the network interface.

$sel:availabilityZone:NetworkInterface', networkInterface_availabilityZone - The Availability Zone.

$sel:privateIpAddress:NetworkInterface', networkInterface_privateIpAddress - The IPv4 address of the network interface within the subnet.

$sel:privateDnsName:NetworkInterface', networkInterface_privateDnsName - The private DNS name.

$sel:requesterId:NetworkInterface', networkInterface_requesterId - The alias or Amazon Web Services account ID of the principal or service that created the network interface.

$sel:description:NetworkInterface', networkInterface_description - A description.

$sel:association:NetworkInterface', networkInterface_association - The association information for an Elastic IP address (IPv4) associated with the network interface.

$sel:ipv6Addresses:NetworkInterface', networkInterface_ipv6Addresses - The IPv6 addresses associated with the network interface.

networkInterface_groups :: Lens' NetworkInterface (Maybe [GroupIdentifier])

Any security groups for the network interface.

networkInterface_status :: Lens' NetworkInterface (Maybe NetworkInterfaceStatus)

The status of the network interface.

networkInterface_privateIpAddresses :: Lens' NetworkInterface (Maybe [NetworkInterfacePrivateIpAddress])

The private IPv4 addresses associated with the network interface.

networkInterface_ipv4Prefixes :: Lens' NetworkInterface (Maybe [Ipv4PrefixSpecification])

The IPv4 prefixes that are assigned to the network interface.

networkInterface_sourceDestCheck :: Lens' NetworkInterface (Maybe Bool)

Indicates whether source/destination checking is enabled.

networkInterface_interfaceType :: Lens' NetworkInterface (Maybe NetworkInterfaceType)

The type of network interface.

networkInterface_vpcId :: Lens' NetworkInterface (Maybe Text)

The ID of the VPC.

networkInterface_tagSet :: Lens' NetworkInterface (Maybe [Tag])

Any tags assigned to the network interface.

networkInterface_requesterManaged :: Lens' NetworkInterface (Maybe Bool)

Indicates whether the network interface is being managed by Amazon Web Services.

networkInterface_outpostArn :: Lens' NetworkInterface (Maybe Text)

The Amazon Resource Name (ARN) of the Outpost.

networkInterface_networkInterfaceId :: Lens' NetworkInterface (Maybe Text)

The ID of the network interface.

networkInterface_subnetId :: Lens' NetworkInterface (Maybe Text)

The ID of the subnet.

networkInterface_macAddress :: Lens' NetworkInterface (Maybe Text)

The MAC address.

networkInterface_attachment :: Lens' NetworkInterface (Maybe NetworkInterfaceAttachment)

The network interface attachment.

networkInterface_ownerId :: Lens' NetworkInterface (Maybe Text)

The Amazon Web Services account ID of the owner of the network interface.

networkInterface_ipv6Prefixes :: Lens' NetworkInterface (Maybe [Ipv6PrefixSpecification])

The IPv6 prefixes that are assigned to the network interface.

networkInterface_availabilityZone :: Lens' NetworkInterface (Maybe Text)

The Availability Zone.

networkInterface_privateIpAddress :: Lens' NetworkInterface (Maybe Text)

The IPv4 address of the network interface within the subnet.

networkInterface_privateDnsName :: Lens' NetworkInterface (Maybe Text)

The private DNS name.

networkInterface_requesterId :: Lens' NetworkInterface (Maybe Text)

The alias or Amazon Web Services account ID of the principal or service that created the network interface.

networkInterface_description :: Lens' NetworkInterface (Maybe Text)

A description.

networkInterface_association :: Lens' NetworkInterface (Maybe NetworkInterfaceAssociation)

The association information for an Elastic IP address (IPv4) associated with the network interface.

networkInterface_ipv6Addresses :: Lens' NetworkInterface (Maybe [NetworkInterfaceIpv6Address])

The IPv6 addresses associated with the network interface.