Amazonka.EC2.Types.OnDemandOptionsRequest

Description

Documentation

data OnDemandOptionsRequest

Describes the configuration of On-Demand Instances in an EC2 Fleet.

See: newOnDemandOptionsRequest smart constructor.

Constructors

OnDemandOptionsRequest'

Fields

capacityReservationOptions :: Maybe CapacityReservationOptionsRequest

The strategy for using unused Capacity Reservations for fulfilling On-Demand capacity. Supported only for fleets of type instant.

singleAvailabilityZone :: Maybe Bool

Indicates that the fleet launches all On-Demand Instances into a single Availability Zone. Supported only for fleets of type instant.

maxTotalPrice :: Maybe Text

The maximum amount per hour for On-Demand Instances that you're willing to pay.

minTargetCapacity :: Maybe Int

The minimum target capacity for On-Demand Instances in the fleet. If the minimum target capacity is not reached, the fleet launches no instances.

singleInstanceType :: Maybe Bool

Indicates that the fleet uses a single instance type to launch all On-Demand Instances in the fleet. Supported only for fleets of type instant.

allocationStrategy :: Maybe FleetOnDemandAllocationStrategy

The order of the launch template overrides to use in fulfilling On-Demand capacity. If you specify lowest-price, EC2 Fleet uses price to determine the order, launching the lowest price first. If you specify prioritized, EC2 Fleet uses the priority that you assigned to each launch template override, launching the highest priority first. If you do not specify a value, EC2 Fleet defaults to lowest-price.

newOnDemandOptionsRequest :: OnDemandOptionsRequest

Create a value of OnDemandOptionsRequest with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:capacityReservationOptions:OnDemandOptionsRequest', onDemandOptionsRequest_capacityReservationOptions - The strategy for using unused Capacity Reservations for fulfilling On-Demand capacity. Supported only for fleets of type instant.

$sel:singleAvailabilityZone:OnDemandOptionsRequest', onDemandOptionsRequest_singleAvailabilityZone - Indicates that the fleet launches all On-Demand Instances into a single Availability Zone. Supported only for fleets of type instant.

$sel:maxTotalPrice:OnDemandOptionsRequest', onDemandOptionsRequest_maxTotalPrice - The maximum amount per hour for On-Demand Instances that you're willing to pay.

$sel:minTargetCapacity:OnDemandOptionsRequest', onDemandOptionsRequest_minTargetCapacity - The minimum target capacity for On-Demand Instances in the fleet. If the minimum target capacity is not reached, the fleet launches no instances.

$sel:singleInstanceType:OnDemandOptionsRequest', onDemandOptionsRequest_singleInstanceType - Indicates that the fleet uses a single instance type to launch all On-Demand Instances in the fleet. Supported only for fleets of type instant.

$sel:allocationStrategy:OnDemandOptionsRequest', onDemandOptionsRequest_allocationStrategy - The order of the launch template overrides to use in fulfilling On-Demand capacity. If you specify lowest-price, EC2 Fleet uses price to determine the order, launching the lowest price first. If you specify prioritized, EC2 Fleet uses the priority that you assigned to each launch template override, launching the highest priority first. If you do not specify a value, EC2 Fleet defaults to lowest-price.

onDemandOptionsRequest_capacityReservationOptions :: Lens' OnDemandOptionsRequest (Maybe CapacityReservationOptionsRequest)

The strategy for using unused Capacity Reservations for fulfilling On-Demand capacity. Supported only for fleets of type instant.

onDemandOptionsRequest_singleAvailabilityZone :: Lens' OnDemandOptionsRequest (Maybe Bool)

Indicates that the fleet launches all On-Demand Instances into a single Availability Zone. Supported only for fleets of type instant.

onDemandOptionsRequest_maxTotalPrice :: Lens' OnDemandOptionsRequest (Maybe Text)

The maximum amount per hour for On-Demand Instances that you're willing to pay.

onDemandOptionsRequest_minTargetCapacity :: Lens' OnDemandOptionsRequest (Maybe Int)

The minimum target capacity for On-Demand Instances in the fleet. If the minimum target capacity is not reached, the fleet launches no instances.

onDemandOptionsRequest_singleInstanceType :: Lens' OnDemandOptionsRequest (Maybe Bool)

Indicates that the fleet uses a single instance type to launch all On-Demand Instances in the fleet. Supported only for fleets of type instant.

onDemandOptionsRequest_allocationStrategy :: Lens' OnDemandOptionsRequest (Maybe FleetOnDemandAllocationStrategy)

The order of the launch template overrides to use in fulfilling On-Demand capacity. If you specify lowest-price, EC2 Fleet uses price to determine the order, launching the lowest price first. If you specify prioritized, EC2 Fleet uses the priority that you assigned to each launch template override, launching the highest priority first. If you do not specify a value, EC2 Fleet defaults to lowest-price.