Amazonka.EC2.Types.RequestSpotLaunchSpecification

Description

Documentation

data RequestSpotLaunchSpecification

Describes the launch specification for an instance.

See: newRequestSpotLaunchSpecification smart constructor.

Constructors

RequestSpotLaunchSpecification'

Fields

securityGroupIds :: Maybe [Text]

One or more security group IDs.

securityGroups :: Maybe [Text]

One or more security groups. When requesting instances in a VPC, you must specify the IDs of the security groups. When requesting instances in EC2-Classic, you can specify the names or the IDs of the security groups.

keyName :: Maybe Text

The name of the key pair.

networkInterfaces :: Maybe [InstanceNetworkInterfaceSpecification]

One or more network interfaces. If you specify a network interface, you must specify subnet IDs and security group IDs using the network interface.

ramdiskId :: Maybe Text

The ID of the RAM disk.

subnetId :: Maybe Text

The ID of the subnet in which to launch the instance.

kernelId :: Maybe Text

The ID of the kernel.

instanceType :: Maybe InstanceType

The instance type.

ebsOptimized :: Maybe Bool

Indicates whether the instance is optimized for EBS I/O. This optimization provides dedicated throughput to Amazon EBS and an optimized configuration stack to provide optimal EBS I/O performance. This optimization isn't available with all instance types. Additional usage charges apply when using an EBS Optimized instance.

Default: false

userData :: Maybe Text

The Base64-encoded user data for the instance. User data is limited to 16 KB.

monitoring :: Maybe RunInstancesMonitoringEnabled

Indicates whether basic or detailed monitoring is enabled for the instance.

Default: Disabled

iamInstanceProfile :: Maybe IamInstanceProfileSpecification

The IAM instance profile.

imageId :: Maybe Text

The ID of the AMI.

addressingType :: Maybe Text

Deprecated.

blockDeviceMappings :: Maybe [BlockDeviceMapping]

One or more block device mapping entries. You can't specify both a snapshot ID and an encryption value. This is because only blank volumes can be encrypted on creation. If a snapshot is the basis for a volume, it is not blank and its encryption status is used for the volume encryption status.

placement :: Maybe SpotPlacement

The placement information for the instance.

newRequestSpotLaunchSpecification :: RequestSpotLaunchSpecification

Create a value of RequestSpotLaunchSpecification with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:securityGroupIds:RequestSpotLaunchSpecification', requestSpotLaunchSpecification_securityGroupIds - One or more security group IDs.

$sel:securityGroups:RequestSpotLaunchSpecification', requestSpotLaunchSpecification_securityGroups - One or more security groups. When requesting instances in a VPC, you must specify the IDs of the security groups. When requesting instances in EC2-Classic, you can specify the names or the IDs of the security groups.

$sel:keyName:RequestSpotLaunchSpecification', requestSpotLaunchSpecification_keyName - The name of the key pair.

$sel:networkInterfaces:RequestSpotLaunchSpecification', requestSpotLaunchSpecification_networkInterfaces - One or more network interfaces. If you specify a network interface, you must specify subnet IDs and security group IDs using the network interface.

$sel:ramdiskId:RequestSpotLaunchSpecification', requestSpotLaunchSpecification_ramdiskId - The ID of the RAM disk.

$sel:subnetId:RequestSpotLaunchSpecification', requestSpotLaunchSpecification_subnetId - The ID of the subnet in which to launch the instance.

$sel:kernelId:RequestSpotLaunchSpecification', requestSpotLaunchSpecification_kernelId - The ID of the kernel.

$sel:instanceType:RequestSpotLaunchSpecification', requestSpotLaunchSpecification_instanceType - The instance type.

$sel:ebsOptimized:RequestSpotLaunchSpecification', requestSpotLaunchSpecification_ebsOptimized - Indicates whether the instance is optimized for EBS I/O. This optimization provides dedicated throughput to Amazon EBS and an optimized configuration stack to provide optimal EBS I/O performance. This optimization isn't available with all instance types. Additional usage charges apply when using an EBS Optimized instance.

Default: false

$sel:userData:RequestSpotLaunchSpecification', requestSpotLaunchSpecification_userData - The Base64-encoded user data for the instance. User data is limited to 16 KB.

$sel:monitoring:RequestSpotLaunchSpecification', requestSpotLaunchSpecification_monitoring - Indicates whether basic or detailed monitoring is enabled for the instance.

Default: Disabled

$sel:iamInstanceProfile:RequestSpotLaunchSpecification', requestSpotLaunchSpecification_iamInstanceProfile - The IAM instance profile.

$sel:imageId:RequestSpotLaunchSpecification', requestSpotLaunchSpecification_imageId - The ID of the AMI.

$sel:addressingType:RequestSpotLaunchSpecification', requestSpotLaunchSpecification_addressingType - Deprecated.

$sel:blockDeviceMappings:RequestSpotLaunchSpecification', requestSpotLaunchSpecification_blockDeviceMappings - One or more block device mapping entries. You can't specify both a snapshot ID and an encryption value. This is because only blank volumes can be encrypted on creation. If a snapshot is the basis for a volume, it is not blank and its encryption status is used for the volume encryption status.

$sel:placement:RequestSpotLaunchSpecification', requestSpotLaunchSpecification_placement - The placement information for the instance.

requestSpotLaunchSpecification_securityGroupIds :: Lens' RequestSpotLaunchSpecification (Maybe [Text])

One or more security group IDs.

requestSpotLaunchSpecification_securityGroups :: Lens' RequestSpotLaunchSpecification (Maybe [Text])

One or more security groups. When requesting instances in a VPC, you must specify the IDs of the security groups. When requesting instances in EC2-Classic, you can specify the names or the IDs of the security groups.

requestSpotLaunchSpecification_keyName :: Lens' RequestSpotLaunchSpecification (Maybe Text)

The name of the key pair.

requestSpotLaunchSpecification_networkInterfaces :: Lens' RequestSpotLaunchSpecification (Maybe [InstanceNetworkInterfaceSpecification])

One or more network interfaces. If you specify a network interface, you must specify subnet IDs and security group IDs using the network interface.

requestSpotLaunchSpecification_ramdiskId :: Lens' RequestSpotLaunchSpecification (Maybe Text)

The ID of the RAM disk.

requestSpotLaunchSpecification_subnetId :: Lens' RequestSpotLaunchSpecification (Maybe Text)

The ID of the subnet in which to launch the instance.

requestSpotLaunchSpecification_kernelId :: Lens' RequestSpotLaunchSpecification (Maybe Text)

The ID of the kernel.

requestSpotLaunchSpecification_instanceType :: Lens' RequestSpotLaunchSpecification (Maybe InstanceType)

The instance type.

requestSpotLaunchSpecification_ebsOptimized :: Lens' RequestSpotLaunchSpecification (Maybe Bool)

Indicates whether the instance is optimized for EBS I/O. This optimization provides dedicated throughput to Amazon EBS and an optimized configuration stack to provide optimal EBS I/O performance. This optimization isn't available with all instance types. Additional usage charges apply when using an EBS Optimized instance.

Default: false

requestSpotLaunchSpecification_userData :: Lens' RequestSpotLaunchSpecification (Maybe Text)

The Base64-encoded user data for the instance. User data is limited to 16 KB.

requestSpotLaunchSpecification_monitoring :: Lens' RequestSpotLaunchSpecification (Maybe RunInstancesMonitoringEnabled)

Indicates whether basic or detailed monitoring is enabled for the instance.

Default: Disabled

requestSpotLaunchSpecification_iamInstanceProfile :: Lens' RequestSpotLaunchSpecification (Maybe IamInstanceProfileSpecification)

The IAM instance profile.

requestSpotLaunchSpecification_imageId :: Lens' RequestSpotLaunchSpecification (Maybe Text)

The ID of the AMI.

requestSpotLaunchSpecification_addressingType :: Lens' RequestSpotLaunchSpecification (Maybe Text)

Deprecated.

requestSpotLaunchSpecification_blockDeviceMappings :: Lens' RequestSpotLaunchSpecification (Maybe [BlockDeviceMapping])

One or more block device mapping entries. You can't specify both a snapshot ID and an encryption value. This is because only blank volumes can be encrypted on creation. If a snapshot is the basis for a volume, it is not blank and its encryption status is used for the volume encryption status.

requestSpotLaunchSpecification_placement :: Lens' RequestSpotLaunchSpecification (Maybe SpotPlacement)

The placement information for the instance.