Amazonka.EC2.Types.ReservedInstancesListing

Description

Documentation

data ReservedInstancesListing

Describes a Reserved Instance listing.

See: newReservedInstancesListing smart constructor.

Constructors

ReservedInstancesListing'

Fields

status :: Maybe ListingStatus

The status of the Reserved Instance listing.

clientToken :: Maybe Text

A unique, case-sensitive key supplied by the client to ensure that the request is idempotent. For more information, see Ensuring Idempotency.

updateDate :: Maybe ISO8601

The last modified timestamp of the listing.

createDate :: Maybe ISO8601

The time the listing was created.

priceSchedules :: Maybe [PriceSchedule]

The price of the Reserved Instance listing.

statusMessage :: Maybe Text

The reason for the current status of the Reserved Instance listing. The response can be blank.

reservedInstancesId :: Maybe Text

The ID of the Reserved Instance.

tags :: Maybe [Tag]

Any tags assigned to the resource.

instanceCounts :: Maybe [InstanceCount]

The number of instances in this state.

reservedInstancesListingId :: Maybe Text

The ID of the Reserved Instance listing.

newReservedInstancesListing :: ReservedInstancesListing

Create a value of ReservedInstancesListing with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:status:ReservedInstancesListing', reservedInstancesListing_status - The status of the Reserved Instance listing.

$sel:clientToken:ReservedInstancesListing', reservedInstancesListing_clientToken - A unique, case-sensitive key supplied by the client to ensure that the request is idempotent. For more information, see Ensuring Idempotency.

$sel:updateDate:ReservedInstancesListing', reservedInstancesListing_updateDate - The last modified timestamp of the listing.

$sel:createDate:ReservedInstancesListing', reservedInstancesListing_createDate - The time the listing was created.

$sel:priceSchedules:ReservedInstancesListing', reservedInstancesListing_priceSchedules - The price of the Reserved Instance listing.

$sel:statusMessage:ReservedInstancesListing', reservedInstancesListing_statusMessage - The reason for the current status of the Reserved Instance listing. The response can be blank.

$sel:reservedInstancesId:ReservedInstancesListing', reservedInstancesListing_reservedInstancesId - The ID of the Reserved Instance.

$sel:tags:ReservedInstancesListing', reservedInstancesListing_tags - Any tags assigned to the resource.

$sel:instanceCounts:ReservedInstancesListing', reservedInstancesListing_instanceCounts - The number of instances in this state.

$sel:reservedInstancesListingId:ReservedInstancesListing', reservedInstancesListing_reservedInstancesListingId - The ID of the Reserved Instance listing.

reservedInstancesListing_status :: Lens' ReservedInstancesListing (Maybe ListingStatus)

The status of the Reserved Instance listing.

reservedInstancesListing_clientToken :: Lens' ReservedInstancesListing (Maybe Text)

A unique, case-sensitive key supplied by the client to ensure that the request is idempotent. For more information, see Ensuring Idempotency.

reservedInstancesListing_updateDate :: Lens' ReservedInstancesListing (Maybe UTCTime)

The last modified timestamp of the listing.

reservedInstancesListing_createDate :: Lens' ReservedInstancesListing (Maybe UTCTime)

The time the listing was created.

reservedInstancesListing_priceSchedules :: Lens' ReservedInstancesListing (Maybe [PriceSchedule])

The price of the Reserved Instance listing.

reservedInstancesListing_statusMessage :: Lens' ReservedInstancesListing (Maybe Text)

The reason for the current status of the Reserved Instance listing. The response can be blank.

reservedInstancesListing_reservedInstancesId :: Lens' ReservedInstancesListing (Maybe Text)

The ID of the Reserved Instance.

reservedInstancesListing_tags :: Lens' ReservedInstancesListing (Maybe [Tag])

Any tags assigned to the resource.

reservedInstancesListing_instanceCounts :: Lens' ReservedInstancesListing (Maybe [InstanceCount])

The number of instances in this state.

reservedInstancesListing_reservedInstancesListingId :: Lens' ReservedInstancesListing (Maybe Text)

The ID of the Reserved Instance listing.