Amazonka.EC2.Types.ReservedInstancesModification

Description

Documentation

data ReservedInstancesModification

Describes a Reserved Instance modification.

See: newReservedInstancesModification smart constructor.

Constructors

ReservedInstancesModification'

Fields

modificationResults :: Maybe [ReservedInstancesModificationResult]

Contains target configurations along with their corresponding new Reserved Instance IDs.

status :: Maybe Text

The status of the Reserved Instances modification request.

clientToken :: Maybe Text

A unique, case-sensitive key supplied by the client to ensure that the request is idempotent. For more information, see Ensuring Idempotency.

updateDate :: Maybe ISO8601

The time when the modification request was last updated.

createDate :: Maybe ISO8601

The time when the modification request was created.

effectiveDate :: Maybe ISO8601

The time for the modification to become effective.

statusMessage :: Maybe Text

The reason for the status.

reservedInstancesModificationId :: Maybe Text

A unique ID for the Reserved Instance modification.

reservedInstancesIds :: Maybe [ReservedInstancesId]

The IDs of one or more Reserved Instances.

newReservedInstancesModification :: ReservedInstancesModification

Create a value of ReservedInstancesModification with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:modificationResults:ReservedInstancesModification', reservedInstancesModification_modificationResults - Contains target configurations along with their corresponding new Reserved Instance IDs.

$sel:status:ReservedInstancesModification', reservedInstancesModification_status - The status of the Reserved Instances modification request.

$sel:clientToken:ReservedInstancesModification', reservedInstancesModification_clientToken - A unique, case-sensitive key supplied by the client to ensure that the request is idempotent. For more information, see Ensuring Idempotency.

$sel:updateDate:ReservedInstancesModification', reservedInstancesModification_updateDate - The time when the modification request was last updated.

$sel:createDate:ReservedInstancesModification', reservedInstancesModification_createDate - The time when the modification request was created.

$sel:effectiveDate:ReservedInstancesModification', reservedInstancesModification_effectiveDate - The time for the modification to become effective.

$sel:statusMessage:ReservedInstancesModification', reservedInstancesModification_statusMessage - The reason for the status.

$sel:reservedInstancesModificationId:ReservedInstancesModification', reservedInstancesModification_reservedInstancesModificationId - A unique ID for the Reserved Instance modification.

$sel:reservedInstancesIds:ReservedInstancesModification', reservedInstancesModification_reservedInstancesIds - The IDs of one or more Reserved Instances.

reservedInstancesModification_modificationResults :: Lens' ReservedInstancesModification (Maybe [ReservedInstancesModificationResult])

Contains target configurations along with their corresponding new Reserved Instance IDs.

reservedInstancesModification_status :: Lens' ReservedInstancesModification (Maybe Text)

The status of the Reserved Instances modification request.

reservedInstancesModification_clientToken :: Lens' ReservedInstancesModification (Maybe Text)

A unique, case-sensitive key supplied by the client to ensure that the request is idempotent. For more information, see Ensuring Idempotency.

reservedInstancesModification_updateDate :: Lens' ReservedInstancesModification (Maybe UTCTime)

The time when the modification request was last updated.

reservedInstancesModification_createDate :: Lens' ReservedInstancesModification (Maybe UTCTime)

The time when the modification request was created.

reservedInstancesModification_effectiveDate :: Lens' ReservedInstancesModification (Maybe UTCTime)

The time for the modification to become effective.

reservedInstancesModification_statusMessage :: Lens' ReservedInstancesModification (Maybe Text)

The reason for the status.

reservedInstancesModification_reservedInstancesModificationId :: Lens' ReservedInstancesModification (Maybe Text)

A unique ID for the Reserved Instance modification.

reservedInstancesModification_reservedInstancesIds :: Lens' ReservedInstancesModification (Maybe [ReservedInstancesId])

The IDs of one or more Reserved Instances.