Amazonka.EC2.Types.SecurityGroupRule

Description

Documentation

data SecurityGroupRule

Describes a security group rule.

See: newSecurityGroupRule smart constructor.

Constructors

SecurityGroupRule'

Fields

cidrIpv4 :: Maybe Text

The IPv4 CIDR range.

isEgress :: Maybe Bool

Indicates whether the security group rule is an outbound rule.

fromPort :: Maybe Int

The start of port range for the TCP and UDP protocols, or an ICMP/ICMPv6 type. A value of -1 indicates all ICMP/ICMPv6 types. If you specify all ICMP/ICMPv6 types, you must specify all codes.

securityGroupRuleId :: Maybe Text

The ID of the security group rule.

groupOwnerId :: Maybe Text

The ID of the Amazon Web Services account that owns the security group.

referencedGroupInfo :: Maybe ReferencedSecurityGroup

Describes the security group that is referenced in the rule.

prefixListId :: Maybe Text

The ID of the prefix list.

ipProtocol :: Maybe Text

The IP protocol name (tcp, udp, icmp, icmpv6) or number (see Protocol Numbers).

Use -1 to specify all protocols.

groupId :: Maybe Text

The ID of the security group.

toPort :: Maybe Int

The end of port range for the TCP and UDP protocols, or an ICMP/ICMPv6 code. A value of -1 indicates all ICMP/ICMPv6 codes. If you specify all ICMP/ICMPv6 types, you must specify all codes.

cidrIpv6 :: Maybe Text

The IPv6 CIDR range.

description :: Maybe Text

The security group rule description.

tags :: Maybe [Tag]

The tags applied to the security group rule.

newSecurityGroupRule :: SecurityGroupRule

Create a value of SecurityGroupRule with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:cidrIpv4:SecurityGroupRule', securityGroupRule_cidrIpv4 - The IPv4 CIDR range.

$sel:isEgress:SecurityGroupRule', securityGroupRule_isEgress - Indicates whether the security group rule is an outbound rule.

$sel:fromPort:SecurityGroupRule', securityGroupRule_fromPort - The start of port range for the TCP and UDP protocols, or an ICMP/ICMPv6 type. A value of -1 indicates all ICMP/ICMPv6 types. If you specify all ICMP/ICMPv6 types, you must specify all codes.

$sel:securityGroupRuleId:SecurityGroupRule', securityGroupRule_securityGroupRuleId - The ID of the security group rule.

$sel:groupOwnerId:SecurityGroupRule', securityGroupRule_groupOwnerId - The ID of the Amazon Web Services account that owns the security group.

$sel:referencedGroupInfo:SecurityGroupRule', securityGroupRule_referencedGroupInfo - Describes the security group that is referenced in the rule.

$sel:prefixListId:SecurityGroupRule', securityGroupRule_prefixListId - The ID of the prefix list.

$sel:ipProtocol:SecurityGroupRule', securityGroupRule_ipProtocol - The IP protocol name (tcp, udp, icmp, icmpv6) or number (see Protocol Numbers).

Use -1 to specify all protocols.

$sel:groupId:SecurityGroupRule', securityGroupRule_groupId - The ID of the security group.

$sel:toPort:SecurityGroupRule', securityGroupRule_toPort - The end of port range for the TCP and UDP protocols, or an ICMP/ICMPv6 code. A value of -1 indicates all ICMP/ICMPv6 codes. If you specify all ICMP/ICMPv6 types, you must specify all codes.

$sel:cidrIpv6:SecurityGroupRule', securityGroupRule_cidrIpv6 - The IPv6 CIDR range.

$sel:description:SecurityGroupRule', securityGroupRule_description - The security group rule description.

$sel:tags:SecurityGroupRule', securityGroupRule_tags - The tags applied to the security group rule.

securityGroupRule_cidrIpv4 :: Lens' SecurityGroupRule (Maybe Text)

The IPv4 CIDR range.

securityGroupRule_isEgress :: Lens' SecurityGroupRule (Maybe Bool)

Indicates whether the security group rule is an outbound rule.

securityGroupRule_fromPort :: Lens' SecurityGroupRule (Maybe Int)

The start of port range for the TCP and UDP protocols, or an ICMP/ICMPv6 type. A value of -1 indicates all ICMP/ICMPv6 types. If you specify all ICMP/ICMPv6 types, you must specify all codes.

securityGroupRule_securityGroupRuleId :: Lens' SecurityGroupRule (Maybe Text)

The ID of the security group rule.

securityGroupRule_groupOwnerId :: Lens' SecurityGroupRule (Maybe Text)

The ID of the Amazon Web Services account that owns the security group.

securityGroupRule_referencedGroupInfo :: Lens' SecurityGroupRule (Maybe ReferencedSecurityGroup)

Describes the security group that is referenced in the rule.

securityGroupRule_prefixListId :: Lens' SecurityGroupRule (Maybe Text)

The ID of the prefix list.

securityGroupRule_ipProtocol :: Lens' SecurityGroupRule (Maybe Text)

The IP protocol name (tcp, udp, icmp, icmpv6) or number (see Protocol Numbers).

Use -1 to specify all protocols.

securityGroupRule_groupId :: Lens' SecurityGroupRule (Maybe Text)

The ID of the security group.

securityGroupRule_toPort :: Lens' SecurityGroupRule (Maybe Int)

The end of port range for the TCP and UDP protocols, or an ICMP/ICMPv6 code. A value of -1 indicates all ICMP/ICMPv6 codes. If you specify all ICMP/ICMPv6 types, you must specify all codes.

securityGroupRule_cidrIpv6 :: Lens' SecurityGroupRule (Maybe Text)

The IPv6 CIDR range.

securityGroupRule_description :: Lens' SecurityGroupRule (Maybe Text)

The security group rule description.

securityGroupRule_tags :: Lens' SecurityGroupRule (Maybe [Tag])

The tags applied to the security group rule.