Amazonka.EC2.Types.TargetCapacitySpecificationRequest

Description

Documentation

data TargetCapacitySpecificationRequest

The number of units to request. You can choose to set the target capacity as the number of instances. Or you can set the target capacity to a performance characteristic that is important to your application workload, such as vCPUs, memory, or I/O. If the request type is maintain, you can specify a target capacity of 0 and add capacity later.

You can use the On-Demand Instance MaxTotalPrice parameter, the Spot Instance MaxTotalPrice parameter, or both parameters to ensure that your fleet cost does not exceed your budget. If you set a maximum price per hour for the On-Demand Instances and Spot Instances in your request, EC2 Fleet will launch instances until it reaches the maximum amount that you're willing to pay. When the maximum amount you're willing to pay is reached, the fleet stops launching instances even if it hasn’t met the target capacity. The MaxTotalPrice parameters are located in OnDemandOptionsRequest and SpotOptionsRequest.

See: newTargetCapacitySpecificationRequest smart constructor.

Constructors

TargetCapacitySpecificationRequest'

Fields

onDemandTargetCapacity :: Maybe Int

The number of On-Demand units to request.

defaultTargetCapacityType :: Maybe DefaultTargetCapacityType

The default TotalTargetCapacity, which is either Spot or On-Demand.

spotTargetCapacity :: Maybe Int

The number of Spot units to request.

totalTargetCapacity :: Int

The number of units to request, filled using DefaultTargetCapacityType.

newTargetCapacitySpecificationRequest

Arguments

:: Int
$sel:totalTargetCapacity:TargetCapacitySpecificationRequest'

-> TargetCapacitySpecificationRequest

Create a value of TargetCapacitySpecificationRequest with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:onDemandTargetCapacity:TargetCapacitySpecificationRequest', targetCapacitySpecificationRequest_onDemandTargetCapacity - The number of On-Demand units to request.

$sel:defaultTargetCapacityType:TargetCapacitySpecificationRequest', targetCapacitySpecificationRequest_defaultTargetCapacityType - The default TotalTargetCapacity, which is either Spot or On-Demand.

$sel:spotTargetCapacity:TargetCapacitySpecificationRequest', targetCapacitySpecificationRequest_spotTargetCapacity - The number of Spot units to request.

$sel:totalTargetCapacity:TargetCapacitySpecificationRequest', targetCapacitySpecificationRequest_totalTargetCapacity - The number of units to request, filled using DefaultTargetCapacityType.

targetCapacitySpecificationRequest_onDemandTargetCapacity :: Lens' TargetCapacitySpecificationRequest (Maybe Int)

The number of On-Demand units to request.

targetCapacitySpecificationRequest_defaultTargetCapacityType :: Lens' TargetCapacitySpecificationRequest (Maybe DefaultTargetCapacityType)

The default TotalTargetCapacity, which is either Spot or On-Demand.

targetCapacitySpecificationRequest_spotTargetCapacity :: Lens' TargetCapacitySpecificationRequest (Maybe Int)

The number of Spot units to request.

targetCapacitySpecificationRequest_totalTargetCapacity :: Lens' TargetCapacitySpecificationRequest Int

The number of units to request, filled using DefaultTargetCapacityType.