Amazonka.EC2.Types.TargetGroupsConfig

Description

Documentation

data TargetGroupsConfig

Describes the target groups to attach to a Spot Fleet. Spot Fleet registers the running Spot Instances with these target groups.

See: newTargetGroupsConfig smart constructor.

Constructors

TargetGroupsConfig'

Fields

targetGroups :: Maybe (NonEmpty TargetGroup)

One or more target groups.

newTargetGroupsConfig :: TargetGroupsConfig

Create a value of TargetGroupsConfig with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:targetGroups:TargetGroupsConfig', targetGroupsConfig_targetGroups - One or more target groups.

targetGroupsConfig_targetGroups :: Lens' TargetGroupsConfig (Maybe (NonEmpty TargetGroup))

One or more target groups.