Amazonka.EC2.Types.TerminateConnectionStatus

Description

Documentation

data TerminateConnectionStatus

Information about a terminated Client VPN endpoint client connection.

See: newTerminateConnectionStatus smart constructor.

Constructors

TerminateConnectionStatus'

Fields

currentStatus :: Maybe ClientVpnConnectionStatus

A message about the status of the client connection, if applicable.

connectionId :: Maybe Text

The ID of the client connection.

previousStatus :: Maybe ClientVpnConnectionStatus

The state of the client connection.

newTerminateConnectionStatus :: TerminateConnectionStatus

Create a value of TerminateConnectionStatus with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:currentStatus:TerminateConnectionStatus', terminateConnectionStatus_currentStatus - A message about the status of the client connection, if applicable.

$sel:connectionId:TerminateConnectionStatus', terminateConnectionStatus_connectionId - The ID of the client connection.

$sel:previousStatus:TerminateConnectionStatus', terminateConnectionStatus_previousStatus - The state of the client connection.

terminateConnectionStatus_currentStatus :: Lens' TerminateConnectionStatus (Maybe ClientVpnConnectionStatus)

A message about the status of the client connection, if applicable.

terminateConnectionStatus_connectionId :: Lens' TerminateConnectionStatus (Maybe Text)

The ID of the client connection.

terminateConnectionStatus_previousStatus :: Lens' TerminateConnectionStatus (Maybe ClientVpnConnectionStatus)

The state of the client connection.