Amazonka.EC2.Types.TrafficMirrorPortRangeRequest

Description

Documentation

data TrafficMirrorPortRangeRequest

Information about the Traffic Mirror filter rule port range.

See: newTrafficMirrorPortRangeRequest smart constructor.

Constructors

TrafficMirrorPortRangeRequest'

Fields

fromPort :: Maybe Int

The first port in the Traffic Mirror port range. This applies to the TCP and UDP protocols.

toPort :: Maybe Int

The last port in the Traffic Mirror port range. This applies to the TCP and UDP protocols.

newTrafficMirrorPortRangeRequest :: TrafficMirrorPortRangeRequest

Create a value of TrafficMirrorPortRangeRequest with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:fromPort:TrafficMirrorPortRangeRequest', trafficMirrorPortRangeRequest_fromPort - The first port in the Traffic Mirror port range. This applies to the TCP and UDP protocols.

$sel:toPort:TrafficMirrorPortRangeRequest', trafficMirrorPortRangeRequest_toPort - The last port in the Traffic Mirror port range. This applies to the TCP and UDP protocols.

trafficMirrorPortRangeRequest_fromPort :: Lens' TrafficMirrorPortRangeRequest (Maybe Int)

The first port in the Traffic Mirror port range. This applies to the TCP and UDP protocols.

trafficMirrorPortRangeRequest_toPort :: Lens' TrafficMirrorPortRangeRequest (Maybe Int)

The last port in the Traffic Mirror port range. This applies to the TCP and UDP protocols.