Amazonka.EC2.Types.TransitGatewayRouteTableAssociation

Description

Documentation

data TransitGatewayRouteTableAssociation

Describes an association between a route table and a resource attachment.

See: newTransitGatewayRouteTableAssociation smart constructor.

Constructors

TransitGatewayRouteTableAssociation'

Fields

state :: Maybe TransitGatewayAssociationState

The state of the association.

resourceId :: Maybe Text

The ID of the resource.

resourceType :: Maybe TransitGatewayAttachmentResourceType

The resource type. Note that the tgw-peering resource type has been deprecated.

transitGatewayAttachmentId :: Maybe Text

The ID of the attachment.

newTransitGatewayRouteTableAssociation :: TransitGatewayRouteTableAssociation

Create a value of TransitGatewayRouteTableAssociation with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:state:TransitGatewayRouteTableAssociation', transitGatewayRouteTableAssociation_state - The state of the association.

$sel:resourceId:TransitGatewayRouteTableAssociation', transitGatewayRouteTableAssociation_resourceId - The ID of the resource.

$sel:resourceType:TransitGatewayRouteTableAssociation', transitGatewayRouteTableAssociation_resourceType - The resource type. Note that the tgw-peering resource type has been deprecated.

$sel:transitGatewayAttachmentId:TransitGatewayRouteTableAssociation', transitGatewayRouteTableAssociation_transitGatewayAttachmentId - The ID of the attachment.

transitGatewayRouteTableAssociation_state :: Lens' TransitGatewayRouteTableAssociation (Maybe TransitGatewayAssociationState)

The state of the association.

transitGatewayRouteTableAssociation_resourceId :: Lens' TransitGatewayRouteTableAssociation (Maybe Text)

The ID of the resource.

transitGatewayRouteTableAssociation_resourceType :: Lens' TransitGatewayRouteTableAssociation (Maybe TransitGatewayAttachmentResourceType)

The resource type. Note that the tgw-peering resource type has been deprecated.

transitGatewayRouteTableAssociation_transitGatewayAttachmentId :: Lens' TransitGatewayRouteTableAssociation (Maybe Text)

The ID of the attachment.