Amazonka.EC2.Types.TrunkInterfaceAssociation

Description

Documentation

data TrunkInterfaceAssociation

Currently available in limited preview only. If you are interested in using this feature, contact your account manager.

Information about an association between a branch network interface with a trunk network interface.

See: newTrunkInterfaceAssociation smart constructor.

Constructors

TrunkInterfaceAssociation'

Fields

associationId :: Maybe Text

The ID of the association.

interfaceProtocol :: Maybe InterfaceProtocolType

The interface protocol. Valid values are VLAN and GRE.

branchInterfaceId :: Maybe Text

The ID of the branch network interface.

greKey :: Maybe Int

The application key when you use the GRE protocol.

vlanId :: Maybe Int

The ID of the VLAN when you use the VLAN protocol.

trunkInterfaceId :: Maybe Text

The ID of the trunk network interface.

tags :: Maybe [Tag]

The tags for the trunk interface association.

newTrunkInterfaceAssociation :: TrunkInterfaceAssociation

Create a value of TrunkInterfaceAssociation with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:associationId:TrunkInterfaceAssociation', trunkInterfaceAssociation_associationId - The ID of the association.

$sel:interfaceProtocol:TrunkInterfaceAssociation', trunkInterfaceAssociation_interfaceProtocol - The interface protocol. Valid values are VLAN and GRE.

$sel:branchInterfaceId:TrunkInterfaceAssociation', trunkInterfaceAssociation_branchInterfaceId - The ID of the branch network interface.

$sel:greKey:TrunkInterfaceAssociation', trunkInterfaceAssociation_greKey - The application key when you use the GRE protocol.

$sel:vlanId:TrunkInterfaceAssociation', trunkInterfaceAssociation_vlanId - The ID of the VLAN when you use the VLAN protocol.

$sel:trunkInterfaceId:TrunkInterfaceAssociation', trunkInterfaceAssociation_trunkInterfaceId - The ID of the trunk network interface.

$sel:tags:TrunkInterfaceAssociation', trunkInterfaceAssociation_tags - The tags for the trunk interface association.

trunkInterfaceAssociation_associationId :: Lens' TrunkInterfaceAssociation (Maybe Text)

The ID of the association.

trunkInterfaceAssociation_interfaceProtocol :: Lens' TrunkInterfaceAssociation (Maybe InterfaceProtocolType)

The interface protocol. Valid values are VLAN and GRE.

trunkInterfaceAssociation_branchInterfaceId :: Lens' TrunkInterfaceAssociation (Maybe Text)

The ID of the branch network interface.

trunkInterfaceAssociation_greKey :: Lens' TrunkInterfaceAssociation (Maybe Int)

The application key when you use the GRE protocol.

trunkInterfaceAssociation_vlanId :: Lens' TrunkInterfaceAssociation (Maybe Int)

The ID of the VLAN when you use the VLAN protocol.

trunkInterfaceAssociation_trunkInterfaceId :: Lens' TrunkInterfaceAssociation (Maybe Text)

The ID of the trunk network interface.

trunkInterfaceAssociation_tags :: Lens' TrunkInterfaceAssociation (Maybe [Tag])

The tags for the trunk interface association.