Amazonka.EC2.Types.TunnelOption

Description

Documentation

data TunnelOption

The VPN tunnel options.

See: newTunnelOption smart constructor.

Constructors

TunnelOption'

Fields

outsideIpAddress :: Maybe Text

The external IP address of the VPN tunnel.

replayWindowSize :: Maybe Int

The number of packets in an IKE replay window.

dpdTimeoutAction :: Maybe Text

The action to take after a DPD timeout occurs.

rekeyFuzzPercentage :: Maybe Int

The percentage of the rekey window determined by RekeyMarginTimeSeconds during which the rekey time is randomly selected.

phase1LifetimeSeconds :: Maybe Int

The lifetime for phase 1 of the IKE negotiation, in seconds.

ikeVersions :: Maybe [IKEVersionsListValue]

The IKE versions that are permitted for the VPN tunnel.

phase2IntegrityAlgorithms :: Maybe [Phase2IntegrityAlgorithmsListValue]

The permitted integrity algorithms for the VPN tunnel for phase 2 IKE negotiations.

phase2LifetimeSeconds :: Maybe Int

The lifetime for phase 2 of the IKE negotiation, in seconds.

phase1EncryptionAlgorithms :: Maybe [Phase1EncryptionAlgorithmsListValue]

The permitted encryption algorithms for the VPN tunnel for phase 1 IKE negotiations.

phase1DHGroupNumbers :: Maybe [Phase1DHGroupNumbersListValue]

The permitted Diffie-Hellman group numbers for the VPN tunnel for phase 1 IKE negotiations.

phase1IntegrityAlgorithms :: Maybe [Phase1IntegrityAlgorithmsListValue]

The permitted integrity algorithms for the VPN tunnel for phase 1 IKE negotiations.

rekeyMarginTimeSeconds :: Maybe Int

The margin time, in seconds, before the phase 2 lifetime expires, during which the Amazon Web Services side of the VPN connection performs an IKE rekey.

dpdTimeoutSeconds :: Maybe Int

The number of seconds after which a DPD timeout occurs.

tunnelInsideCidr :: Maybe Text

The range of inside IPv4 addresses for the tunnel.

startupAction :: Maybe Text

The action to take when the establishing the VPN tunnels for a VPN connection.

phase2EncryptionAlgorithms :: Maybe [Phase2EncryptionAlgorithmsListValue]

The permitted encryption algorithms for the VPN tunnel for phase 2 IKE negotiations.

phase2DHGroupNumbers :: Maybe [Phase2DHGroupNumbersListValue]

The permitted Diffie-Hellman group numbers for the VPN tunnel for phase 2 IKE negotiations.

preSharedKey :: Maybe Text

The pre-shared key (PSK) to establish initial authentication between the virtual private gateway and the customer gateway.

tunnelInsideIpv6Cidr :: Maybe Text

The range of inside IPv6 addresses for the tunnel.

newTunnelOption :: TunnelOption

Create a value of TunnelOption with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:outsideIpAddress:TunnelOption', tunnelOption_outsideIpAddress - The external IP address of the VPN tunnel.

$sel:replayWindowSize:TunnelOption', tunnelOption_replayWindowSize - The number of packets in an IKE replay window.

$sel:dpdTimeoutAction:TunnelOption', tunnelOption_dpdTimeoutAction - The action to take after a DPD timeout occurs.

$sel:rekeyFuzzPercentage:TunnelOption', tunnelOption_rekeyFuzzPercentage - The percentage of the rekey window determined by RekeyMarginTimeSeconds during which the rekey time is randomly selected.

$sel:phase1LifetimeSeconds:TunnelOption', tunnelOption_phase1LifetimeSeconds - The lifetime for phase 1 of the IKE negotiation, in seconds.

$sel:ikeVersions:TunnelOption', tunnelOption_ikeVersions - The IKE versions that are permitted for the VPN tunnel.

$sel:phase2IntegrityAlgorithms:TunnelOption', tunnelOption_phase2IntegrityAlgorithms - The permitted integrity algorithms for the VPN tunnel for phase 2 IKE negotiations.

$sel:phase2LifetimeSeconds:TunnelOption', tunnelOption_phase2LifetimeSeconds - The lifetime for phase 2 of the IKE negotiation, in seconds.

$sel:phase1EncryptionAlgorithms:TunnelOption', tunnelOption_phase1EncryptionAlgorithms - The permitted encryption algorithms for the VPN tunnel for phase 1 IKE negotiations.

$sel:phase1DHGroupNumbers:TunnelOption', tunnelOption_phase1DHGroupNumbers - The permitted Diffie-Hellman group numbers for the VPN tunnel for phase 1 IKE negotiations.

$sel:phase1IntegrityAlgorithms:TunnelOption', tunnelOption_phase1IntegrityAlgorithms - The permitted integrity algorithms for the VPN tunnel for phase 1 IKE negotiations.

$sel:rekeyMarginTimeSeconds:TunnelOption', tunnelOption_rekeyMarginTimeSeconds - The margin time, in seconds, before the phase 2 lifetime expires, during which the Amazon Web Services side of the VPN connection performs an IKE rekey.

$sel:dpdTimeoutSeconds:TunnelOption', tunnelOption_dpdTimeoutSeconds - The number of seconds after which a DPD timeout occurs.

$sel:tunnelInsideCidr:TunnelOption', tunnelOption_tunnelInsideCidr - The range of inside IPv4 addresses for the tunnel.

$sel:startupAction:TunnelOption', tunnelOption_startupAction - The action to take when the establishing the VPN tunnels for a VPN connection.

$sel:phase2EncryptionAlgorithms:TunnelOption', tunnelOption_phase2EncryptionAlgorithms - The permitted encryption algorithms for the VPN tunnel for phase 2 IKE negotiations.

$sel:phase2DHGroupNumbers:TunnelOption', tunnelOption_phase2DHGroupNumbers - The permitted Diffie-Hellman group numbers for the VPN tunnel for phase 2 IKE negotiations.

$sel:preSharedKey:TunnelOption', tunnelOption_preSharedKey - The pre-shared key (PSK) to establish initial authentication between the virtual private gateway and the customer gateway.

$sel:tunnelInsideIpv6Cidr:TunnelOption', tunnelOption_tunnelInsideIpv6Cidr - The range of inside IPv6 addresses for the tunnel.

tunnelOption_outsideIpAddress :: Lens' TunnelOption (Maybe Text)

The external IP address of the VPN tunnel.

tunnelOption_replayWindowSize :: Lens' TunnelOption (Maybe Int)

The number of packets in an IKE replay window.

tunnelOption_dpdTimeoutAction :: Lens' TunnelOption (Maybe Text)

The action to take after a DPD timeout occurs.

tunnelOption_rekeyFuzzPercentage :: Lens' TunnelOption (Maybe Int)

The percentage of the rekey window determined by RekeyMarginTimeSeconds during which the rekey time is randomly selected.

tunnelOption_phase1LifetimeSeconds :: Lens' TunnelOption (Maybe Int)

The lifetime for phase 1 of the IKE negotiation, in seconds.

tunnelOption_ikeVersions :: Lens' TunnelOption (Maybe [IKEVersionsListValue])

The IKE versions that are permitted for the VPN tunnel.

tunnelOption_phase2IntegrityAlgorithms :: Lens' TunnelOption (Maybe [Phase2IntegrityAlgorithmsListValue])

The permitted integrity algorithms for the VPN tunnel for phase 2 IKE negotiations.

tunnelOption_phase2LifetimeSeconds :: Lens' TunnelOption (Maybe Int)

The lifetime for phase 2 of the IKE negotiation, in seconds.

tunnelOption_phase1EncryptionAlgorithms :: Lens' TunnelOption (Maybe [Phase1EncryptionAlgorithmsListValue])

The permitted encryption algorithms for the VPN tunnel for phase 1 IKE negotiations.

tunnelOption_phase1DHGroupNumbers :: Lens' TunnelOption (Maybe [Phase1DHGroupNumbersListValue])

The permitted Diffie-Hellman group numbers for the VPN tunnel for phase 1 IKE negotiations.

tunnelOption_phase1IntegrityAlgorithms :: Lens' TunnelOption (Maybe [Phase1IntegrityAlgorithmsListValue])

The permitted integrity algorithms for the VPN tunnel for phase 1 IKE negotiations.

tunnelOption_rekeyMarginTimeSeconds :: Lens' TunnelOption (Maybe Int)

The margin time, in seconds, before the phase 2 lifetime expires, during which the Amazon Web Services side of the VPN connection performs an IKE rekey.

tunnelOption_dpdTimeoutSeconds :: Lens' TunnelOption (Maybe Int)

The number of seconds after which a DPD timeout occurs.

tunnelOption_tunnelInsideCidr :: Lens' TunnelOption (Maybe Text)

The range of inside IPv4 addresses for the tunnel.

tunnelOption_startupAction :: Lens' TunnelOption (Maybe Text)

The action to take when the establishing the VPN tunnels for a VPN connection.

tunnelOption_phase2EncryptionAlgorithms :: Lens' TunnelOption (Maybe [Phase2EncryptionAlgorithmsListValue])

The permitted encryption algorithms for the VPN tunnel for phase 2 IKE negotiations.

tunnelOption_phase2DHGroupNumbers :: Lens' TunnelOption (Maybe [Phase2DHGroupNumbersListValue])

The permitted Diffie-Hellman group numbers for the VPN tunnel for phase 2 IKE negotiations.

tunnelOption_preSharedKey :: Lens' TunnelOption (Maybe Text)

The pre-shared key (PSK) to establish initial authentication between the virtual private gateway and the customer gateway.

tunnelOption_tunnelInsideIpv6Cidr :: Lens' TunnelOption (Maybe Text)

The range of inside IPv6 addresses for the tunnel.