Amazonka.EC2.Types.Volume

Description

Documentation

data Volume

Describes a volume.

See: newVolume smart constructor.

Constructors

Volume'

Fields

fastRestored :: Maybe Bool

Indicates whether the volume was created using fast snapshot restore.

multiAttachEnabled :: Maybe Bool

Indicates whether Amazon EBS Multi-Attach is enabled.

attachments :: Maybe [VolumeAttachment]

Information about the volume attachments.

throughput :: Maybe Int

The throughput that the volume supports, in MiB/s.

iops :: Maybe Int

The number of I/O operations per second (IOPS). For gp3, io1, and io2 volumes, this represents the number of IOPS that are provisioned for the volume. For gp2 volumes, this represents the baseline performance of the volume and the rate at which the volume accumulates I/O credits for bursting.

outpostArn :: Maybe Text

The Amazon Resource Name (ARN) of the Outpost.

kmsKeyId :: Maybe Text

The Amazon Resource Name (ARN) of the Key Management Service (KMS) KMS key that was used to protect the volume encryption key for the volume.

tags :: Maybe [Tag]

Any tags assigned to the volume.

availabilityZone :: Text

The Availability Zone for the volume.

createTime :: ISO8601

The time stamp when volume creation was initiated.

encrypted :: Bool

Indicates whether the volume is encrypted.

size :: Int

The size of the volume, in GiBs.

snapshotId :: Text

The snapshot from which the volume was created, if applicable.

state :: VolumeState

The volume state.

volumeId :: Text

The ID of the volume.

volumeType :: VolumeType

The volume type.

newVolume

Arguments

:: Text
$sel:availabilityZone:Volume'

-> UTCTime
$sel:createTime:Volume'

-> Bool
$sel:encrypted:Volume'

-> Int
$sel:size:Volume'

-> Text
$sel:snapshotId:Volume'

-> VolumeState
$sel:state:Volume'

-> Text
$sel:volumeId:Volume'

-> VolumeType
$sel:volumeType:Volume'

-> Volume

Create a value of Volume with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:fastRestored:Volume', volume_fastRestored - Indicates whether the volume was created using fast snapshot restore.

$sel:multiAttachEnabled:Volume', volume_multiAttachEnabled - Indicates whether Amazon EBS Multi-Attach is enabled.

$sel:attachments:Volume', volume_attachments - Information about the volume attachments.

$sel:throughput:Volume', volume_throughput - The throughput that the volume supports, in MiB/s.

$sel:iops:Volume', volume_iops - The number of I/O operations per second (IOPS). For gp3, io1, and io2 volumes, this represents the number of IOPS that are provisioned for the volume. For gp2 volumes, this represents the baseline performance of the volume and the rate at which the volume accumulates I/O credits for bursting.

$sel:outpostArn:Volume', volume_outpostArn - The Amazon Resource Name (ARN) of the Outpost.

$sel:kmsKeyId:Volume', volume_kmsKeyId - The Amazon Resource Name (ARN) of the Key Management Service (KMS) KMS key that was used to protect the volume encryption key for the volume.

$sel:tags:Volume', volume_tags - Any tags assigned to the volume.

$sel:availabilityZone:Volume', volume_availabilityZone - The Availability Zone for the volume.

$sel:createTime:Volume', volume_createTime - The time stamp when volume creation was initiated.

$sel:encrypted:Volume', volume_encrypted - Indicates whether the volume is encrypted.

$sel:size:Volume', volume_size - The size of the volume, in GiBs.

$sel:snapshotId:Volume', volume_snapshotId - The snapshot from which the volume was created, if applicable.

$sel:state:Volume', volume_state - The volume state.

$sel:volumeId:Volume', volume_volumeId - The ID of the volume.

$sel:volumeType:Volume', volume_volumeType - The volume type.

volume_fastRestored :: Lens' Volume (Maybe Bool)

Indicates whether the volume was created using fast snapshot restore.

volume_multiAttachEnabled :: Lens' Volume (Maybe Bool)

Indicates whether Amazon EBS Multi-Attach is enabled.

volume_attachments :: Lens' Volume (Maybe [VolumeAttachment])

Information about the volume attachments.

volume_throughput :: Lens' Volume (Maybe Int)

The throughput that the volume supports, in MiB/s.

volume_iops :: Lens' Volume (Maybe Int)

The number of I/O operations per second (IOPS). For gp3, io1, and io2 volumes, this represents the number of IOPS that are provisioned for the volume. For gp2 volumes, this represents the baseline performance of the volume and the rate at which the volume accumulates I/O credits for bursting.

volume_outpostArn :: Lens' Volume (Maybe Text)

The Amazon Resource Name (ARN) of the Outpost.

volume_kmsKeyId :: Lens' Volume (Maybe Text)

The Amazon Resource Name (ARN) of the Key Management Service (KMS) KMS key that was used to protect the volume encryption key for the volume.

volume_tags :: Lens' Volume (Maybe [Tag])

Any tags assigned to the volume.

volume_availabilityZone :: Lens' Volume Text

The Availability Zone for the volume.

volume_createTime :: Lens' Volume UTCTime

The time stamp when volume creation was initiated.

volume_encrypted :: Lens' Volume Bool

Indicates whether the volume is encrypted.

volume_size :: Lens' Volume Int

The size of the volume, in GiBs.

volume_snapshotId :: Lens' Volume Text

The snapshot from which the volume was created, if applicable.

volume_state :: Lens' Volume VolumeState

The volume state.

volume_volumeId :: Lens' Volume Text

The ID of the volume.

volume_volumeType :: Lens' Volume VolumeType

The volume type.