Amazonka.EC2.Types.VolumeStatusItem

Description

Documentation

data VolumeStatusItem

Describes the volume status.

See: newVolumeStatusItem smart constructor.

Constructors

VolumeStatusItem'

Fields

volumeStatus :: Maybe VolumeStatusInfo

The volume status.

actions :: Maybe [VolumeStatusAction]

The details of the operation.

outpostArn :: Maybe Text

The Amazon Resource Name (ARN) of the Outpost.

events :: Maybe [VolumeStatusEvent]

A list of events associated with the volume.

availabilityZone :: Maybe Text

The Availability Zone of the volume.

volumeId :: Maybe Text

The volume ID.

attachmentStatuses :: Maybe [VolumeStatusAttachmentStatus]

Information about the instances to which the volume is attached.

newVolumeStatusItem :: VolumeStatusItem

Create a value of VolumeStatusItem with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:volumeStatus:VolumeStatusItem', volumeStatusItem_volumeStatus - The volume status.

$sel:actions:VolumeStatusItem', volumeStatusItem_actions - The details of the operation.

$sel:outpostArn:VolumeStatusItem', volumeStatusItem_outpostArn - The Amazon Resource Name (ARN) of the Outpost.

$sel:events:VolumeStatusItem', volumeStatusItem_events - A list of events associated with the volume.

$sel:availabilityZone:VolumeStatusItem', volumeStatusItem_availabilityZone - The Availability Zone of the volume.

$sel:volumeId:VolumeStatusItem', volumeStatusItem_volumeId - The volume ID.

$sel:attachmentStatuses:VolumeStatusItem', volumeStatusItem_attachmentStatuses - Information about the instances to which the volume is attached.

volumeStatusItem_volumeStatus :: Lens' VolumeStatusItem (Maybe VolumeStatusInfo)

The volume status.

volumeStatusItem_actions :: Lens' VolumeStatusItem (Maybe [VolumeStatusAction])

The details of the operation.

volumeStatusItem_outpostArn :: Lens' VolumeStatusItem (Maybe Text)

The Amazon Resource Name (ARN) of the Outpost.

volumeStatusItem_events :: Lens' VolumeStatusItem (Maybe [VolumeStatusEvent])

A list of events associated with the volume.

volumeStatusItem_availabilityZone :: Lens' VolumeStatusItem (Maybe Text)

The Availability Zone of the volume.

volumeStatusItem_volumeId :: Lens' VolumeStatusItem (Maybe Text)

The volume ID.

volumeStatusItem_attachmentStatuses :: Lens' VolumeStatusItem (Maybe [VolumeStatusAttachmentStatus])

Information about the instances to which the volume is attached.