Amazonka.EC2InstanceConnect.Lens

Description

Operations

SendSSHPublicKey

sendSSHPublicKey_instanceId :: Lens' SendSSHPublicKey Text

The ID of the EC2 instance.

sendSSHPublicKey_instanceOSUser :: Lens' SendSSHPublicKey Text

The OS user on the EC2 instance for whom the key can be used to authenticate.

sendSSHPublicKey_sSHPublicKey :: Lens' SendSSHPublicKey Text

The public key material. To use the public key, you must have the matching private key.

sendSSHPublicKey_availabilityZone :: Lens' SendSSHPublicKey Text

The Availability Zone in which the EC2 instance was launched.

sendSSHPublicKeyResponse_requestId :: Lens' SendSSHPublicKeyResponse (Maybe Text)

The ID of the request. Please provide this ID when contacting AWS Support for assistance.

sendSSHPublicKeyResponse_success :: Lens' SendSSHPublicKeyResponse (Maybe Bool)

Is true if the request succeeds and an error otherwise.

sendSSHPublicKeyResponse_httpStatus :: Lens' SendSSHPublicKeyResponse Int

The response's http status code.

SendSerialConsoleSSHPublicKey

sendSerialConsoleSSHPublicKey_serialPort :: Lens' SendSerialConsoleSSHPublicKey (Maybe Natural)

The serial port of the EC2 instance. Currently only port 0 is supported.

Default: 0

sendSerialConsoleSSHPublicKey_instanceId :: Lens' SendSerialConsoleSSHPublicKey Text

The ID of the EC2 instance.

sendSerialConsoleSSHPublicKey_sSHPublicKey :: Lens' SendSerialConsoleSSHPublicKey Text

The public key material. To use the public key, you must have the matching private key. For information about the supported key formats and lengths, see Requirements for key pairs in the Amazon EC2 User Guide.

sendSerialConsoleSSHPublicKeyResponse_requestId :: Lens' SendSerialConsoleSSHPublicKeyResponse (Maybe Text)

The ID of the request. Please provide this ID when contacting AWS Support for assistance.

sendSerialConsoleSSHPublicKeyResponse_success :: Lens' SendSerialConsoleSSHPublicKeyResponse (Maybe Bool)

Is true if the request succeeds and an error otherwise.

sendSerialConsoleSSHPublicKeyResponse_httpStatus :: Lens' SendSerialConsoleSSHPublicKeyResponse Int

The response's http status code.

Types