Amazonka.EC2InstanceConnect.SendSerialConsoleSSHPublicKey

Description

Pushes an SSH public key to the specified EC2 instance. The key remains for 60 seconds, which gives you 60 seconds to establish a serial console connection to the instance using SSH. For more information, see EC2 Serial Console in the Amazon EC2 User Guide.

Creating a Request

data SendSerialConsoleSSHPublicKey

See: newSendSerialConsoleSSHPublicKey smart constructor.

Constructors

SendSerialConsoleSSHPublicKey'

Fields

serialPort :: Maybe Natural

The serial port of the EC2 instance. Currently only port 0 is supported.

Default: 0

instanceId :: Text

The ID of the EC2 instance.

sSHPublicKey :: Text

The public key material. To use the public key, you must have the matching private key. For information about the supported key formats and lengths, see Requirements for key pairs in the Amazon EC2 User Guide.

newSendSerialConsoleSSHPublicKey

Arguments

:: Text
$sel:instanceId:SendSerialConsoleSSHPublicKey'

-> Text
$sel:sSHPublicKey:SendSerialConsoleSSHPublicKey'

-> SendSerialConsoleSSHPublicKey

Create a value of SendSerialConsoleSSHPublicKey with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:serialPort:SendSerialConsoleSSHPublicKey', sendSerialConsoleSSHPublicKey_serialPort - The serial port of the EC2 instance. Currently only port 0 is supported.

Default: 0

$sel:instanceId:SendSerialConsoleSSHPublicKey', sendSerialConsoleSSHPublicKey_instanceId - The ID of the EC2 instance.

$sel:sSHPublicKey:SendSerialConsoleSSHPublicKey', sendSerialConsoleSSHPublicKey_sSHPublicKey - The public key material. To use the public key, you must have the matching private key. For information about the supported key formats and lengths, see Requirements for key pairs in the Amazon EC2 User Guide.

Request Lenses

sendSerialConsoleSSHPublicKey_serialPort :: Lens' SendSerialConsoleSSHPublicKey (Maybe Natural)

The serial port of the EC2 instance. Currently only port 0 is supported.

Default: 0

sendSerialConsoleSSHPublicKey_instanceId :: Lens' SendSerialConsoleSSHPublicKey Text

The ID of the EC2 instance.

sendSerialConsoleSSHPublicKey_sSHPublicKey :: Lens' SendSerialConsoleSSHPublicKey Text

The public key material. To use the public key, you must have the matching private key. For information about the supported key formats and lengths, see Requirements for key pairs in the Amazon EC2 User Guide.

Destructuring the Response

data SendSerialConsoleSSHPublicKeyResponse

See: newSendSerialConsoleSSHPublicKeyResponse smart constructor.

Constructors

SendSerialConsoleSSHPublicKeyResponse'

Fields

requestId :: Maybe Text

The ID of the request. Please provide this ID when contacting AWS Support for assistance.

success :: Maybe Bool

Is true if the request succeeds and an error otherwise.

httpStatus :: Int

The response's http status code.

newSendSerialConsoleSSHPublicKeyResponse

Arguments

:: Int
$sel:httpStatus:SendSerialConsoleSSHPublicKeyResponse'

-> SendSerialConsoleSSHPublicKeyResponse

Create a value of SendSerialConsoleSSHPublicKeyResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:requestId:SendSerialConsoleSSHPublicKeyResponse', sendSerialConsoleSSHPublicKeyResponse_requestId - The ID of the request. Please provide this ID when contacting AWS Support for assistance.

$sel:success:SendSerialConsoleSSHPublicKeyResponse', sendSerialConsoleSSHPublicKeyResponse_success - Is true if the request succeeds and an error otherwise.

$sel:httpStatus:SendSerialConsoleSSHPublicKeyResponse', sendSerialConsoleSSHPublicKeyResponse_httpStatus - The response's http status code.

Response Lenses

sendSerialConsoleSSHPublicKeyResponse_requestId :: Lens' SendSerialConsoleSSHPublicKeyResponse (Maybe Text)

The ID of the request. Please provide this ID when contacting AWS Support for assistance.

sendSerialConsoleSSHPublicKeyResponse_success :: Lens' SendSerialConsoleSSHPublicKeyResponse (Maybe Bool)

Is true if the request succeeds and an error otherwise.

sendSerialConsoleSSHPublicKeyResponse_httpStatus :: Lens' SendSerialConsoleSSHPublicKeyResponse Int

The response's http status code.