Amazonka.ECR.BatchCheckLayerAvailability

Description

Checks the availability of one or more image layers in a repository.

When an image is pushed to a repository, each image layer is checked to verify if it has been uploaded before. If it has been uploaded, then the image layer is skipped.

This operation is used by the Amazon ECR proxy and is not generally used by customers for pulling and pushing images. In most cases, you should use the docker CLI to pull, tag, and push images.

Creating a Request

data BatchCheckLayerAvailability

See: newBatchCheckLayerAvailability smart constructor.

Constructors

BatchCheckLayerAvailability'

Fields

registryId :: Maybe Text

The Amazon Web Services account ID associated with the registry that contains the image layers to check. If you do not specify a registry, the default registry is assumed.

repositoryName :: Text

The name of the repository that is associated with the image layers to check.

layerDigests :: NonEmpty Text

The digests of the image layers to check.

newBatchCheckLayerAvailability

Arguments

:: Text
$sel:repositoryName:BatchCheckLayerAvailability'

-> NonEmpty Text
$sel:layerDigests:BatchCheckLayerAvailability'

-> BatchCheckLayerAvailability

Create a value of BatchCheckLayerAvailability with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:registryId:BatchCheckLayerAvailability', batchCheckLayerAvailability_registryId - The Amazon Web Services account ID associated with the registry that contains the image layers to check. If you do not specify a registry, the default registry is assumed.

$sel:repositoryName:BatchCheckLayerAvailability', batchCheckLayerAvailability_repositoryName - The name of the repository that is associated with the image layers to check.

$sel:layerDigests:BatchCheckLayerAvailability', batchCheckLayerAvailability_layerDigests - The digests of the image layers to check.

Request Lenses

batchCheckLayerAvailability_registryId :: Lens' BatchCheckLayerAvailability (Maybe Text)

The Amazon Web Services account ID associated with the registry that contains the image layers to check. If you do not specify a registry, the default registry is assumed.

batchCheckLayerAvailability_repositoryName :: Lens' BatchCheckLayerAvailability Text

The name of the repository that is associated with the image layers to check.

batchCheckLayerAvailability_layerDigests :: Lens' BatchCheckLayerAvailability (NonEmpty Text)

The digests of the image layers to check.

Destructuring the Response

data BatchCheckLayerAvailabilityResponse

See: newBatchCheckLayerAvailabilityResponse smart constructor.

Constructors

BatchCheckLayerAvailabilityResponse'

Fields

failures :: Maybe [LayerFailure]

Any failures associated with the call.

layers :: Maybe [Layer]

A list of image layer objects corresponding to the image layer references in the request.

httpStatus :: Int

The response's http status code.

newBatchCheckLayerAvailabilityResponse

Arguments

:: Int
$sel:httpStatus:BatchCheckLayerAvailabilityResponse'

-> BatchCheckLayerAvailabilityResponse

Create a value of BatchCheckLayerAvailabilityResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:failures:BatchCheckLayerAvailabilityResponse', batchCheckLayerAvailabilityResponse_failures - Any failures associated with the call.

$sel:layers:BatchCheckLayerAvailabilityResponse', batchCheckLayerAvailabilityResponse_layers - A list of image layer objects corresponding to the image layer references in the request.

$sel:httpStatus:BatchCheckLayerAvailabilityResponse', batchCheckLayerAvailabilityResponse_httpStatus - The response's http status code.

Response Lenses

batchCheckLayerAvailabilityResponse_failures :: Lens' BatchCheckLayerAvailabilityResponse (Maybe [LayerFailure])

Any failures associated with the call.

batchCheckLayerAvailabilityResponse_layers :: Lens' BatchCheckLayerAvailabilityResponse (Maybe [Layer])

A list of image layer objects corresponding to the image layer references in the request.

batchCheckLayerAvailabilityResponse_httpStatus :: Lens' BatchCheckLayerAvailabilityResponse Int

The response's http status code.