Amazonka.ECR.DescribeRepositories

Description

Describes image repositories in a registry.

This operation returns paginated results.

Creating a Request

data DescribeRepositories

See: newDescribeRepositories smart constructor.

Constructors

DescribeRepositories'

Fields

registryId :: Maybe Text

The Amazon Web Services account ID associated with the registry that contains the repositories to be described. If you do not specify a registry, the default registry is assumed.

repositoryNames :: Maybe (NonEmpty Text)

A list of repositories to describe. If this parameter is omitted, then all repositories in a registry are described.

nextToken :: Maybe Text

The nextToken value returned from a previous paginated DescribeRepositories request where maxResults was used and the results exceeded the value of that parameter. Pagination continues from the end of the previous results that returned the nextToken value. This value is null when there are no more results to return. This option cannot be used when you specify repositories with repositoryNames.

This token should be treated as an opaque identifier that is only used to retrieve the next items in a list and not for other programmatic purposes.

maxResults :: Maybe Natural

The maximum number of repository results returned by DescribeRepositories in paginated output. When this parameter is used, DescribeRepositories only returns maxResults results in a single page along with a nextToken response element. The remaining results of the initial request can be seen by sending another DescribeRepositories request with the returned nextToken value. This value can be between 1 and 1000. If this parameter is not used, then DescribeRepositories returns up to 100 results and a nextToken value, if applicable. This option cannot be used when you specify repositories with repositoryNames.

newDescribeRepositories :: DescribeRepositories

Create a value of DescribeRepositories with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:registryId:DescribeRepositories', describeRepositories_registryId - The Amazon Web Services account ID associated with the registry that contains the repositories to be described. If you do not specify a registry, the default registry is assumed.

$sel:repositoryNames:DescribeRepositories', describeRepositories_repositoryNames - A list of repositories to describe. If this parameter is omitted, then all repositories in a registry are described.

$sel:nextToken:DescribeRepositories', describeRepositories_nextToken - The nextToken value returned from a previous paginated DescribeRepositories request where maxResults was used and the results exceeded the value of that parameter. Pagination continues from the end of the previous results that returned the nextToken value. This value is null when there are no more results to return. This option cannot be used when you specify repositories with repositoryNames.

This token should be treated as an opaque identifier that is only used to retrieve the next items in a list and not for other programmatic purposes.

$sel:maxResults:DescribeRepositories', describeRepositories_maxResults - The maximum number of repository results returned by DescribeRepositories in paginated output. When this parameter is used, DescribeRepositories only returns maxResults results in a single page along with a nextToken response element. The remaining results of the initial request can be seen by sending another DescribeRepositories request with the returned nextToken value. This value can be between 1 and 1000. If this parameter is not used, then DescribeRepositories returns up to 100 results and a nextToken value, if applicable. This option cannot be used when you specify repositories with repositoryNames.

Request Lenses

describeRepositories_registryId :: Lens' DescribeRepositories (Maybe Text)

The Amazon Web Services account ID associated with the registry that contains the repositories to be described. If you do not specify a registry, the default registry is assumed.

describeRepositories_repositoryNames :: Lens' DescribeRepositories (Maybe (NonEmpty Text))

A list of repositories to describe. If this parameter is omitted, then all repositories in a registry are described.

describeRepositories_nextToken :: Lens' DescribeRepositories (Maybe Text)

The nextToken value returned from a previous paginated DescribeRepositories request where maxResults was used and the results exceeded the value of that parameter. Pagination continues from the end of the previous results that returned the nextToken value. This value is null when there are no more results to return. This option cannot be used when you specify repositories with repositoryNames.

This token should be treated as an opaque identifier that is only used to retrieve the next items in a list and not for other programmatic purposes.

describeRepositories_maxResults :: Lens' DescribeRepositories (Maybe Natural)

The maximum number of repository results returned by DescribeRepositories in paginated output. When this parameter is used, DescribeRepositories only returns maxResults results in a single page along with a nextToken response element. The remaining results of the initial request can be seen by sending another DescribeRepositories request with the returned nextToken value. This value can be between 1 and 1000. If this parameter is not used, then DescribeRepositories returns up to 100 results and a nextToken value, if applicable. This option cannot be used when you specify repositories with repositoryNames.

Destructuring the Response

data DescribeRepositoriesResponse

See: newDescribeRepositoriesResponse smart constructor.

Constructors

DescribeRepositoriesResponse'

Fields

repositories :: Maybe [Repository]

A list of repository objects corresponding to valid repositories.

nextToken :: Maybe Text

The nextToken value to include in a future DescribeRepositories request. When the results of a DescribeRepositories request exceed maxResults, this value can be used to retrieve the next page of results. This value is null when there are no more results to return.

httpStatus :: Int

The response's http status code.

newDescribeRepositoriesResponse

Arguments

:: Int
$sel:httpStatus:DescribeRepositoriesResponse'

-> DescribeRepositoriesResponse

Create a value of DescribeRepositoriesResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:repositories:DescribeRepositoriesResponse', describeRepositoriesResponse_repositories - A list of repository objects corresponding to valid repositories.

$sel:nextToken:DescribeRepositories', describeRepositoriesResponse_nextToken - The nextToken value to include in a future DescribeRepositories request. When the results of a DescribeRepositories request exceed maxResults, this value can be used to retrieve the next page of results. This value is null when there are no more results to return.

$sel:httpStatus:DescribeRepositoriesResponse', describeRepositoriesResponse_httpStatus - The response's http status code.

Response Lenses

describeRepositoriesResponse_repositories :: Lens' DescribeRepositoriesResponse (Maybe [Repository])

A list of repository objects corresponding to valid repositories.

describeRepositoriesResponse_nextToken :: Lens' DescribeRepositoriesResponse (Maybe Text)

The nextToken value to include in a future DescribeRepositories request. When the results of a DescribeRepositories request exceed maxResults, this value can be used to retrieve the next page of results. This value is null when there are no more results to return.

describeRepositoriesResponse_httpStatus :: Lens' DescribeRepositoriesResponse Int

The response's http status code.