Amazonka.ECR.GetLifecyclePolicyPreview

Description

Retrieves the results of the lifecycle policy preview request for the specified repository.

This operation returns paginated results.

Creating a Request

data GetLifecyclePolicyPreview

See: newGetLifecyclePolicyPreview smart constructor.

Constructors

GetLifecyclePolicyPreview'

Fields

registryId :: Maybe Text

The Amazon Web Services account ID associated with the registry that contains the repository. If you do not specify a registry, the default registry is assumed.

imageIds :: Maybe [ImageIdentifier]

The list of imageIDs to be included.

nextToken :: Maybe Text

The nextToken value returned from a previous paginated GetLifecyclePolicyPreviewRequest request where maxResults was used and the results exceeded the value of that parameter. Pagination continues from the end of the previous results that returned the nextToken value. This value is null when there are no more results to return. This option cannot be used when you specify images with imageIds.

filter' :: Maybe LifecyclePolicyPreviewFilter

An optional parameter that filters results based on image tag status and all tags, if tagged.

maxResults :: Maybe Natural

The maximum number of repository results returned by GetLifecyclePolicyPreviewRequest in paginated output. When this parameter is used, GetLifecyclePolicyPreviewRequest only returns maxResults results in a single page along with a nextToken response element. The remaining results of the initial request can be seen by sending another GetLifecyclePolicyPreviewRequest request with the returned nextToken value. This value can be between 1 and 1000. If this parameter is not used, then GetLifecyclePolicyPreviewRequest returns up to 100 results and a nextToken value, if applicable. This option cannot be used when you specify images with imageIds.

repositoryName :: Text

The name of the repository.

newGetLifecyclePolicyPreview

Arguments

:: Text
$sel:repositoryName:GetLifecyclePolicyPreview'

-> GetLifecyclePolicyPreview

Create a value of GetLifecyclePolicyPreview with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:registryId:GetLifecyclePolicyPreview', getLifecyclePolicyPreview_registryId - The Amazon Web Services account ID associated with the registry that contains the repository. If you do not specify a registry, the default registry is assumed.

$sel:imageIds:GetLifecyclePolicyPreview', getLifecyclePolicyPreview_imageIds - The list of imageIDs to be included.

$sel:nextToken:GetLifecyclePolicyPreview', getLifecyclePolicyPreview_nextToken - The nextToken value returned from a previous paginated GetLifecyclePolicyPreviewRequest request where maxResults was used and the results exceeded the value of that parameter. Pagination continues from the end of the previous results that returned the nextToken value. This value is null when there are no more results to return. This option cannot be used when you specify images with imageIds.

$sel:filter':GetLifecyclePolicyPreview', getLifecyclePolicyPreview_filter - An optional parameter that filters results based on image tag status and all tags, if tagged.

$sel:maxResults:GetLifecyclePolicyPreview', getLifecyclePolicyPreview_maxResults - The maximum number of repository results returned by GetLifecyclePolicyPreviewRequest in paginated output. When this parameter is used, GetLifecyclePolicyPreviewRequest only returns maxResults results in a single page along with a nextToken response element. The remaining results of the initial request can be seen by sending another GetLifecyclePolicyPreviewRequest request with the returned nextToken value. This value can be between 1 and 1000. If this parameter is not used, then GetLifecyclePolicyPreviewRequest returns up to 100 results and a nextToken value, if applicable. This option cannot be used when you specify images with imageIds.

$sel:repositoryName:GetLifecyclePolicyPreview', getLifecyclePolicyPreview_repositoryName - The name of the repository.

Request Lenses

getLifecyclePolicyPreview_registryId :: Lens' GetLifecyclePolicyPreview (Maybe Text)

The Amazon Web Services account ID associated with the registry that contains the repository. If you do not specify a registry, the default registry is assumed.

getLifecyclePolicyPreview_imageIds :: Lens' GetLifecyclePolicyPreview (Maybe [ImageIdentifier])

The list of imageIDs to be included.

getLifecyclePolicyPreview_nextToken :: Lens' GetLifecyclePolicyPreview (Maybe Text)

The nextToken value returned from a previous paginated GetLifecyclePolicyPreviewRequest request where maxResults was used and the results exceeded the value of that parameter. Pagination continues from the end of the previous results that returned the nextToken value. This value is null when there are no more results to return. This option cannot be used when you specify images with imageIds.

getLifecyclePolicyPreview_filter :: Lens' GetLifecyclePolicyPreview (Maybe LifecyclePolicyPreviewFilter)

An optional parameter that filters results based on image tag status and all tags, if tagged.

getLifecyclePolicyPreview_maxResults :: Lens' GetLifecyclePolicyPreview (Maybe Natural)

The maximum number of repository results returned by GetLifecyclePolicyPreviewRequest in paginated output. When this parameter is used, GetLifecyclePolicyPreviewRequest only returns maxResults results in a single page along with a nextToken response element. The remaining results of the initial request can be seen by sending another GetLifecyclePolicyPreviewRequest request with the returned nextToken value. This value can be between 1 and 1000. If this parameter is not used, then GetLifecyclePolicyPreviewRequest returns up to 100 results and a nextToken value, if applicable. This option cannot be used when you specify images with imageIds.

getLifecyclePolicyPreview_repositoryName :: Lens' GetLifecyclePolicyPreview Text

The name of the repository.

Destructuring the Response

data GetLifecyclePolicyPreviewResponse

See: newGetLifecyclePolicyPreviewResponse smart constructor.

Constructors

GetLifecyclePolicyPreviewResponse'

Fields

summary :: Maybe LifecyclePolicyPreviewSummary

The list of images that is returned as a result of the action.

status :: Maybe LifecyclePolicyPreviewStatus

The status of the lifecycle policy preview request.

registryId :: Maybe Text

The registry ID associated with the request.

lifecyclePolicyText :: Maybe Text

The JSON lifecycle policy text.

nextToken :: Maybe Text

The nextToken value to include in a future GetLifecyclePolicyPreview request. When the results of a GetLifecyclePolicyPreview request exceed maxResults, this value can be used to retrieve the next page of results. This value is null when there are no more results to return.

repositoryName :: Maybe Text

The repository name associated with the request.

previewResults :: Maybe [LifecyclePolicyPreviewResult]

The results of the lifecycle policy preview request.

httpStatus :: Int

The response's http status code.

newGetLifecyclePolicyPreviewResponse

Arguments

:: Int
$sel:httpStatus:GetLifecyclePolicyPreviewResponse'

-> GetLifecyclePolicyPreviewResponse

Create a value of GetLifecyclePolicyPreviewResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:summary:GetLifecyclePolicyPreviewResponse', getLifecyclePolicyPreviewResponse_summary - The list of images that is returned as a result of the action.

$sel:status:GetLifecyclePolicyPreviewResponse', getLifecyclePolicyPreviewResponse_status - The status of the lifecycle policy preview request.

$sel:registryId:GetLifecyclePolicyPreview', getLifecyclePolicyPreviewResponse_registryId - The registry ID associated with the request.

$sel:lifecyclePolicyText:GetLifecyclePolicyPreviewResponse', getLifecyclePolicyPreviewResponse_lifecyclePolicyText - The JSON lifecycle policy text.

$sel:nextToken:GetLifecyclePolicyPreview', getLifecyclePolicyPreviewResponse_nextToken - The nextToken value to include in a future GetLifecyclePolicyPreview request. When the results of a GetLifecyclePolicyPreview request exceed maxResults, this value can be used to retrieve the next page of results. This value is null when there are no more results to return.

$sel:repositoryName:GetLifecyclePolicyPreview', getLifecyclePolicyPreviewResponse_repositoryName - The repository name associated with the request.

$sel:previewResults:GetLifecyclePolicyPreviewResponse', getLifecyclePolicyPreviewResponse_previewResults - The results of the lifecycle policy preview request.

$sel:httpStatus:GetLifecyclePolicyPreviewResponse', getLifecyclePolicyPreviewResponse_httpStatus - The response's http status code.

Response Lenses

getLifecyclePolicyPreviewResponse_summary :: Lens' GetLifecyclePolicyPreviewResponse (Maybe LifecyclePolicyPreviewSummary)

The list of images that is returned as a result of the action.

getLifecyclePolicyPreviewResponse_status :: Lens' GetLifecyclePolicyPreviewResponse (Maybe LifecyclePolicyPreviewStatus)

The status of the lifecycle policy preview request.

getLifecyclePolicyPreviewResponse_registryId :: Lens' GetLifecyclePolicyPreviewResponse (Maybe Text)

The registry ID associated with the request.

getLifecyclePolicyPreviewResponse_lifecyclePolicyText :: Lens' GetLifecyclePolicyPreviewResponse (Maybe Text)

The JSON lifecycle policy text.

getLifecyclePolicyPreviewResponse_nextToken :: Lens' GetLifecyclePolicyPreviewResponse (Maybe Text)

The nextToken value to include in a future GetLifecyclePolicyPreview request. When the results of a GetLifecyclePolicyPreview request exceed maxResults, this value can be used to retrieve the next page of results. This value is null when there are no more results to return.

getLifecyclePolicyPreviewResponse_repositoryName :: Lens' GetLifecyclePolicyPreviewResponse (Maybe Text)

The repository name associated with the request.

getLifecyclePolicyPreviewResponse_previewResults :: Lens' GetLifecyclePolicyPreviewResponse (Maybe [LifecyclePolicyPreviewResult])

The results of the lifecycle policy preview request.

getLifecyclePolicyPreviewResponse_httpStatus :: Lens' GetLifecyclePolicyPreviewResponse Int

The response's http status code.