Amazonka.ECR.PutImageTagMutability

Description

Updates the image tag mutability settings for the specified repository. For more information, see Image tag mutability in the Amazon Elastic Container Registry User Guide.

Creating a Request

data PutImageTagMutability

See: newPutImageTagMutability smart constructor.

Constructors

PutImageTagMutability'

Fields

registryId :: Maybe Text

The Amazon Web Services account ID associated with the registry that contains the repository in which to update the image tag mutability settings. If you do not specify a registry, the default registry is assumed.

repositoryName :: Text

The name of the repository in which to update the image tag mutability settings.

imageTagMutability :: ImageTagMutability

The tag mutability setting for the repository. If MUTABLE is specified, image tags can be overwritten. If IMMUTABLE is specified, all image tags within the repository will be immutable which will prevent them from being overwritten.

newPutImageTagMutability

Arguments

:: Text
$sel:repositoryName:PutImageTagMutability'

-> ImageTagMutability
$sel:imageTagMutability:PutImageTagMutability'

-> PutImageTagMutability

Create a value of PutImageTagMutability with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:registryId:PutImageTagMutability', putImageTagMutability_registryId - The Amazon Web Services account ID associated with the registry that contains the repository in which to update the image tag mutability settings. If you do not specify a registry, the default registry is assumed.

$sel:repositoryName:PutImageTagMutability', putImageTagMutability_repositoryName - The name of the repository in which to update the image tag mutability settings.

$sel:imageTagMutability:PutImageTagMutability', putImageTagMutability_imageTagMutability - The tag mutability setting for the repository. If MUTABLE is specified, image tags can be overwritten. If IMMUTABLE is specified, all image tags within the repository will be immutable which will prevent them from being overwritten.

Request Lenses

putImageTagMutability_registryId :: Lens' PutImageTagMutability (Maybe Text)

The Amazon Web Services account ID associated with the registry that contains the repository in which to update the image tag mutability settings. If you do not specify a registry, the default registry is assumed.

putImageTagMutability_repositoryName :: Lens' PutImageTagMutability Text

The name of the repository in which to update the image tag mutability settings.

putImageTagMutability_imageTagMutability :: Lens' PutImageTagMutability ImageTagMutability

The tag mutability setting for the repository. If MUTABLE is specified, image tags can be overwritten. If IMMUTABLE is specified, all image tags within the repository will be immutable which will prevent them from being overwritten.

Destructuring the Response

data PutImageTagMutabilityResponse

See: newPutImageTagMutabilityResponse smart constructor.

Constructors

PutImageTagMutabilityResponse'

Fields

registryId :: Maybe Text

The registry ID associated with the request.

repositoryName :: Maybe Text

The repository name associated with the request.

imageTagMutability :: Maybe ImageTagMutability

The image tag mutability setting for the repository.

httpStatus :: Int

The response's http status code.

newPutImageTagMutabilityResponse

Arguments

:: Int
$sel:httpStatus:PutImageTagMutabilityResponse'

-> PutImageTagMutabilityResponse

Create a value of PutImageTagMutabilityResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:registryId:PutImageTagMutability', putImageTagMutabilityResponse_registryId - The registry ID associated with the request.

$sel:repositoryName:PutImageTagMutability', putImageTagMutabilityResponse_repositoryName - The repository name associated with the request.

$sel:imageTagMutability:PutImageTagMutability', putImageTagMutabilityResponse_imageTagMutability - The image tag mutability setting for the repository.

$sel:httpStatus:PutImageTagMutabilityResponse', putImageTagMutabilityResponse_httpStatus - The response's http status code.

Response Lenses

putImageTagMutabilityResponse_registryId :: Lens' PutImageTagMutabilityResponse (Maybe Text)

The registry ID associated with the request.

putImageTagMutabilityResponse_repositoryName :: Lens' PutImageTagMutabilityResponse (Maybe Text)

The repository name associated with the request.

putImageTagMutabilityResponse_imageTagMutability :: Lens' PutImageTagMutabilityResponse (Maybe ImageTagMutability)

The image tag mutability setting for the repository.

putImageTagMutabilityResponse_httpStatus :: Lens' PutImageTagMutabilityResponse Int

The response's http status code.