Amazonka.ECR.StartImageScan

Description

Starts an image vulnerability scan. An image scan can only be started once per 24 hours on an individual image. This limit includes if an image was scanned on initial push. For more information, see Image scanning in the Amazon Elastic Container Registry User Guide.

Creating a Request

data StartImageScan

See: newStartImageScan smart constructor.

Constructors

StartImageScan'

Fields

registryId :: Maybe Text

The Amazon Web Services account ID associated with the registry that contains the repository in which to start an image scan request. If you do not specify a registry, the default registry is assumed.

repositoryName :: Text

The name of the repository that contains the images to scan.

imageId :: ImageIdentifier

newStartImageScan

Arguments

:: Text
$sel:repositoryName:StartImageScan'

-> ImageIdentifier
$sel:imageId:StartImageScan'

-> StartImageScan

Create a value of StartImageScan with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:registryId:StartImageScan', startImageScan_registryId - The Amazon Web Services account ID associated with the registry that contains the repository in which to start an image scan request. If you do not specify a registry, the default registry is assumed.

$sel:repositoryName:StartImageScan', startImageScan_repositoryName - The name of the repository that contains the images to scan.

$sel:imageId:StartImageScan', startImageScan_imageId - Undocumented member.

Request Lenses

startImageScan_registryId :: Lens' StartImageScan (Maybe Text)

The Amazon Web Services account ID associated with the registry that contains the repository in which to start an image scan request. If you do not specify a registry, the default registry is assumed.

startImageScan_repositoryName :: Lens' StartImageScan Text

The name of the repository that contains the images to scan.

startImageScan_imageId :: Lens' StartImageScan ImageIdentifier

Undocumented member.

Destructuring the Response

data StartImageScanResponse

See: newStartImageScanResponse smart constructor.

Constructors

StartImageScanResponse'

Fields

registryId :: Maybe Text

The registry ID associated with the request.

imageScanStatus :: Maybe ImageScanStatus

The current state of the scan.

imageId :: Maybe ImageIdentifier

repositoryName :: Maybe Text

The repository name associated with the request.

httpStatus :: Int

The response's http status code.

newStartImageScanResponse

Arguments

:: Int
$sel:httpStatus:StartImageScanResponse'

-> StartImageScanResponse

Create a value of StartImageScanResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:registryId:StartImageScan', startImageScanResponse_registryId - The registry ID associated with the request.

$sel:imageScanStatus:StartImageScanResponse', startImageScanResponse_imageScanStatus - The current state of the scan.

$sel:imageId:StartImageScan', startImageScanResponse_imageId - Undocumented member.

$sel:repositoryName:StartImageScan', startImageScanResponse_repositoryName - The repository name associated with the request.

$sel:httpStatus:StartImageScanResponse', startImageScanResponse_httpStatus - The response's http status code.

Response Lenses

startImageScanResponse_registryId :: Lens' StartImageScanResponse (Maybe Text)

The registry ID associated with the request.

startImageScanResponse_imageScanStatus :: Lens' StartImageScanResponse (Maybe ImageScanStatus)

The current state of the scan.

startImageScanResponse_imageId :: Lens' StartImageScanResponse (Maybe ImageIdentifier)

Undocumented member.

startImageScanResponse_repositoryName :: Lens' StartImageScanResponse (Maybe Text)

The repository name associated with the request.

startImageScanResponse_httpStatus :: Lens' StartImageScanResponse Int

The response's http status code.