Amazonka.ECR.Types.ImageDetail

Description

Documentation

data ImageDetail

An object that describes an image returned by a DescribeImages operation.

See: newImageDetail smart constructor.

Constructors

ImageDetail'

Fields

registryId :: Maybe Text

The Amazon Web Services account ID associated with the registry to which this image belongs.

imageTags :: Maybe [Text]

The list of tags associated with this image.

imageScanStatus :: Maybe ImageScanStatus

The current state of the scan.

imageManifestMediaType :: Maybe Text

The media type of the image manifest.

imageSizeInBytes :: Maybe Integer

The size, in bytes, of the image in the repository.

If the image is a manifest list, this will be the max size of all manifests in the list.

Beginning with Docker version 1.9, the Docker client compresses image layers before pushing them to a V2 Docker registry. The output of the docker images command shows the uncompressed image size, so it may return a larger image size than the image sizes returned by DescribeImages.

imageDigest :: Maybe Text

The sha256 digest of the image manifest.

imageScanFindingsSummary :: Maybe ImageScanFindingsSummary

A summary of the last completed image scan.

artifactMediaType :: Maybe Text

The artifact media type of the image.

imagePushedAt :: Maybe POSIX

The date and time, expressed in standard JavaScript date format, at which the current image was pushed to the repository.

repositoryName :: Maybe Text

The name of the repository to which this image belongs.

newImageDetail :: ImageDetail

Create a value of ImageDetail with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:registryId:ImageDetail', imageDetail_registryId - The Amazon Web Services account ID associated with the registry to which this image belongs.

$sel:imageTags:ImageDetail', imageDetail_imageTags - The list of tags associated with this image.

$sel:imageScanStatus:ImageDetail', imageDetail_imageScanStatus - The current state of the scan.

$sel:imageManifestMediaType:ImageDetail', imageDetail_imageManifestMediaType - The media type of the image manifest.

$sel:imageSizeInBytes:ImageDetail', imageDetail_imageSizeInBytes - The size, in bytes, of the image in the repository.

If the image is a manifest list, this will be the max size of all manifests in the list.

Beginning with Docker version 1.9, the Docker client compresses image layers before pushing them to a V2 Docker registry. The output of the docker images command shows the uncompressed image size, so it may return a larger image size than the image sizes returned by DescribeImages.

$sel:imageDigest:ImageDetail', imageDetail_imageDigest - The sha256 digest of the image manifest.

$sel:imageScanFindingsSummary:ImageDetail', imageDetail_imageScanFindingsSummary - A summary of the last completed image scan.

$sel:artifactMediaType:ImageDetail', imageDetail_artifactMediaType - The artifact media type of the image.

$sel:imagePushedAt:ImageDetail', imageDetail_imagePushedAt - The date and time, expressed in standard JavaScript date format, at which the current image was pushed to the repository.

$sel:repositoryName:ImageDetail', imageDetail_repositoryName - The name of the repository to which this image belongs.

imageDetail_registryId :: Lens' ImageDetail (Maybe Text)

The Amazon Web Services account ID associated with the registry to which this image belongs.

imageDetail_imageTags :: Lens' ImageDetail (Maybe [Text])

The list of tags associated with this image.

imageDetail_imageScanStatus :: Lens' ImageDetail (Maybe ImageScanStatus)

The current state of the scan.

imageDetail_imageManifestMediaType :: Lens' ImageDetail (Maybe Text)

The media type of the image manifest.

imageDetail_imageSizeInBytes :: Lens' ImageDetail (Maybe Integer)

The size, in bytes, of the image in the repository.

If the image is a manifest list, this will be the max size of all manifests in the list.

Beginning with Docker version 1.9, the Docker client compresses image layers before pushing them to a V2 Docker registry. The output of the docker images command shows the uncompressed image size, so it may return a larger image size than the image sizes returned by DescribeImages.

imageDetail_imageDigest :: Lens' ImageDetail (Maybe Text)

The sha256 digest of the image manifest.

imageDetail_imageScanFindingsSummary :: Lens' ImageDetail (Maybe ImageScanFindingsSummary)

A summary of the last completed image scan.

imageDetail_artifactMediaType :: Lens' ImageDetail (Maybe Text)

The artifact media type of the image.

imageDetail_imagePushedAt :: Lens' ImageDetail (Maybe UTCTime)

The date and time, expressed in standard JavaScript date format, at which the current image was pushed to the repository.

imageDetail_repositoryName :: Lens' ImageDetail (Maybe Text)

The name of the repository to which this image belongs.