Amazonka.ECRPublic.Lens

Description

Operations

GetRepositoryPolicy

getRepositoryPolicy_registryId :: Lens' GetRepositoryPolicy (Maybe Text)

The AWS account ID associated with the public registry that contains the repository. If you do not specify a registry, the default public registry is assumed.

getRepositoryPolicy_repositoryName :: Lens' GetRepositoryPolicy Text

The name of the repository with the policy to retrieve.

getRepositoryPolicyResponse_registryId :: Lens' GetRepositoryPolicyResponse (Maybe Text)

The registry ID associated with the request.

getRepositoryPolicyResponse_repositoryName :: Lens' GetRepositoryPolicyResponse (Maybe Text)

The repository name associated with the request.

getRepositoryPolicyResponse_policyText :: Lens' GetRepositoryPolicyResponse (Maybe Text)

The repository policy text associated with the repository. The policy text will be in JSON format.

getRepositoryPolicyResponse_httpStatus :: Lens' GetRepositoryPolicyResponse Int

The response's http status code.

PutRegistryCatalogData

putRegistryCatalogData_displayName :: Lens' PutRegistryCatalogData (Maybe Text)

The display name for a public registry. The display name is shown as the repository author in the Amazon ECR Public Gallery.

The registry display name is only publicly visible in the Amazon ECR Public Gallery for verified accounts.

putRegistryCatalogDataResponse_httpStatus :: Lens' PutRegistryCatalogDataResponse Int

The response's http status code.

putRegistryCatalogDataResponse_registryCatalogData :: Lens' PutRegistryCatalogDataResponse RegistryCatalogData

The catalog data for the public registry.

BatchDeleteImage

batchDeleteImage_registryId :: Lens' BatchDeleteImage (Maybe Text)

The AWS account ID associated with the registry that contains the image to delete. If you do not specify a registry, the default public registry is assumed.

batchDeleteImage_repositoryName :: Lens' BatchDeleteImage Text

The repository in a public registry that contains the image to delete.

batchDeleteImage_imageIds :: Lens' BatchDeleteImage (NonEmpty ImageIdentifier)

A list of image ID references that correspond to images to delete. The format of the imageIds reference is imageTag=tag or imageDigest=digest.

batchDeleteImageResponse_failures :: Lens' BatchDeleteImageResponse (Maybe [ImageFailure])

Any failures associated with the call.

batchDeleteImageResponse_imageIds :: Lens' BatchDeleteImageResponse (Maybe (NonEmpty ImageIdentifier))

The image IDs of the deleted images.

batchDeleteImageResponse_httpStatus :: Lens' BatchDeleteImageResponse Int

The response's http status code.

ListTagsForResource

listTagsForResource_resourceArn :: Lens' ListTagsForResource Text

The Amazon Resource Name (ARN) that identifies the resource for which to list the tags. Currently, the supported resource is an Amazon ECR Public repository.

listTagsForResourceResponse_tags :: Lens' ListTagsForResourceResponse (Maybe [Tag])

The tags for the resource.

listTagsForResourceResponse_httpStatus :: Lens' ListTagsForResourceResponse Int

The response's http status code.

BatchCheckLayerAvailability

batchCheckLayerAvailability_registryId :: Lens' BatchCheckLayerAvailability (Maybe Text)

The AWS account ID associated with the public registry that contains the image layers to check. If you do not specify a registry, the default public registry is assumed.

batchCheckLayerAvailability_repositoryName :: Lens' BatchCheckLayerAvailability Text

The name of the repository that is associated with the image layers to check.

batchCheckLayerAvailability_layerDigests :: Lens' BatchCheckLayerAvailability (NonEmpty Text)

The digests of the image layers to check.

batchCheckLayerAvailabilityResponse_failures :: Lens' BatchCheckLayerAvailabilityResponse (Maybe [LayerFailure])

Any failures associated with the call.

batchCheckLayerAvailabilityResponse_layers :: Lens' BatchCheckLayerAvailabilityResponse (Maybe [Layer])

A list of image layer objects corresponding to the image layer references in the request.

batchCheckLayerAvailabilityResponse_httpStatus :: Lens' BatchCheckLayerAvailabilityResponse Int

The response's http status code.

PutRepositoryCatalogData

putRepositoryCatalogData_registryId :: Lens' PutRepositoryCatalogData (Maybe Text)

The AWS account ID associated with the public registry the repository is in. If you do not specify a registry, the default public registry is assumed.

putRepositoryCatalogData_repositoryName :: Lens' PutRepositoryCatalogData Text

The name of the repository to create or update the catalog data for.

putRepositoryCatalogData_catalogData :: Lens' PutRepositoryCatalogData RepositoryCatalogDataInput

An object containing the catalog data for a repository. This data is publicly visible in the Amazon ECR Public Gallery.

putRepositoryCatalogDataResponse_catalogData :: Lens' PutRepositoryCatalogDataResponse (Maybe RepositoryCatalogData)

The catalog data for the repository.

putRepositoryCatalogDataResponse_httpStatus :: Lens' PutRepositoryCatalogDataResponse Int

The response's http status code.

DeleteRepositoryPolicy

deleteRepositoryPolicy_registryId :: Lens' DeleteRepositoryPolicy (Maybe Text)

The AWS account ID associated with the public registry that contains the repository policy to delete. If you do not specify a registry, the default public registry is assumed.

deleteRepositoryPolicy_repositoryName :: Lens' DeleteRepositoryPolicy Text

The name of the repository that is associated with the repository policy to delete.

deleteRepositoryPolicyResponse_registryId :: Lens' DeleteRepositoryPolicyResponse (Maybe Text)

The registry ID associated with the request.

deleteRepositoryPolicyResponse_repositoryName :: Lens' DeleteRepositoryPolicyResponse (Maybe Text)

The repository name associated with the request.

deleteRepositoryPolicyResponse_policyText :: Lens' DeleteRepositoryPolicyResponse (Maybe Text)

The JSON repository policy that was deleted from the repository.

deleteRepositoryPolicyResponse_httpStatus :: Lens' DeleteRepositoryPolicyResponse Int

The response's http status code.

CreateRepository

createRepository_catalogData :: Lens' CreateRepository (Maybe RepositoryCatalogDataInput)

The details about the repository that are publicly visible in the Amazon ECR Public Gallery.

createRepository_tags :: Lens' CreateRepository (Maybe [Tag])

The metadata that you apply to the repository to help you categorize and organize them. Each tag consists of a key and an optional value, both of which you define. Tag keys can have a maximum character length of 128 characters, and tag values can have a maximum length of 256 characters.

createRepository_repositoryName :: Lens' CreateRepository Text

The name to use for the repository. This appears publicly in the Amazon ECR Public Gallery. The repository name may be specified on its own (such as nginx-web-app) or it can be prepended with a namespace to group the repository into a category (such as project-a/nginx-web-app).

createRepositoryResponse_repository :: Lens' CreateRepositoryResponse (Maybe Repository)

The repository that was created.

createRepositoryResponse_catalogData :: Lens' CreateRepositoryResponse (Maybe RepositoryCatalogData)

Undocumented member.

createRepositoryResponse_httpStatus :: Lens' CreateRepositoryResponse Int

The response's http status code.

DescribeRegistries

describeRegistries_nextToken :: Lens' DescribeRegistries (Maybe Text)

The nextToken value returned from a previous paginated DescribeRegistries request where maxResults was used and the results exceeded the value of that parameter. Pagination continues from the end of the previous results that returned the nextToken value. This value is null when there are no more results to return.

This token should be treated as an opaque identifier that is only used to retrieve the next items in a list and not for other programmatic purposes.

describeRegistries_maxResults :: Lens' DescribeRegistries (Maybe Natural)

The maximum number of repository results returned by DescribeRegistries in paginated output. When this parameter is used, DescribeRegistries only returns maxResults results in a single page along with a nextToken response element. The remaining results of the initial request can be seen by sending another DescribeRegistries request with the returned nextToken value. This value can be between 1 and 1000. If this parameter is not used, then DescribeRegistries returns up to 100 results and a nextToken value, if applicable.

describeRegistriesResponse_nextToken :: Lens' DescribeRegistriesResponse (Maybe Text)

The nextToken value to include in a future DescribeRepositories request. When the results of a DescribeRepositories request exceed maxResults, this value can be used to retrieve the next page of results. This value is null when there are no more results to return.

describeRegistriesResponse_httpStatus :: Lens' DescribeRegistriesResponse Int

The response's http status code.

describeRegistriesResponse_registries :: Lens' DescribeRegistriesResponse [Registry]

An object containing the details for a public registry.

CompleteLayerUpload

completeLayerUpload_registryId :: Lens' CompleteLayerUpload (Maybe Text)

The AWS account ID associated with the registry to which to upload layers. If you do not specify a registry, the default public registry is assumed.

completeLayerUpload_repositoryName :: Lens' CompleteLayerUpload Text

The name of the repository in a public registry to associate with the image layer.

completeLayerUpload_uploadId :: Lens' CompleteLayerUpload Text

The upload ID from a previous InitiateLayerUpload operation to associate with the image layer.

completeLayerUpload_layerDigests :: Lens' CompleteLayerUpload (NonEmpty Text)

The sha256 digest of the image layer.

completeLayerUploadResponse_registryId :: Lens' CompleteLayerUploadResponse (Maybe Text)

The public registry ID associated with the request.

completeLayerUploadResponse_layerDigest :: Lens' CompleteLayerUploadResponse (Maybe Text)

The sha256 digest of the image layer.

completeLayerUploadResponse_repositoryName :: Lens' CompleteLayerUploadResponse (Maybe Text)

The repository name associated with the request.

completeLayerUploadResponse_uploadId :: Lens' CompleteLayerUploadResponse (Maybe Text)

The upload ID associated with the layer.

completeLayerUploadResponse_httpStatus :: Lens' CompleteLayerUploadResponse Int

The response's http status code.

DescribeRepositories

describeRepositories_registryId :: Lens' DescribeRepositories (Maybe Text)

The AWS account ID associated with the registry that contains the repositories to be described. If you do not specify a registry, the default public registry is assumed.

describeRepositories_repositoryNames :: Lens' DescribeRepositories (Maybe (NonEmpty Text))

A list of repositories to describe. If this parameter is omitted, then all repositories in a registry are described.

describeRepositories_nextToken :: Lens' DescribeRepositories (Maybe Text)

The nextToken value returned from a previous paginated DescribeRepositories request where maxResults was used and the results exceeded the value of that parameter. Pagination continues from the end of the previous results that returned the nextToken value. This value is null when there are no more results to return. This option cannot be used when you specify repositories with repositoryNames.

This token should be treated as an opaque identifier that is only used to retrieve the next items in a list and not for other programmatic purposes.

describeRepositories_maxResults :: Lens' DescribeRepositories (Maybe Natural)

The maximum number of repository results returned by DescribeRepositories in paginated output. When this parameter is used, DescribeRepositories only returns maxResults results in a single page along with a nextToken response element. The remaining results of the initial request can be seen by sending another DescribeRepositories request with the returned nextToken value. This value can be between 1 and 1000. If this parameter is not used, then DescribeRepositories returns up to 100 results and a nextToken value, if applicable. This option cannot be used when you specify repositories with repositoryNames.

describeRepositoriesResponse_repositories :: Lens' DescribeRepositoriesResponse (Maybe [Repository])

A list of repository objects corresponding to valid repositories.

describeRepositoriesResponse_nextToken :: Lens' DescribeRepositoriesResponse (Maybe Text)

The nextToken value to include in a future DescribeRepositories request. When the results of a DescribeRepositories request exceed maxResults, this value can be used to retrieve the next page of results. This value is null when there are no more results to return.

describeRepositoriesResponse_httpStatus :: Lens' DescribeRepositoriesResponse Int

The response's http status code.

UploadLayerPart

uploadLayerPart_registryId :: Lens' UploadLayerPart (Maybe Text)

The AWS account ID associated with the registry to which you are uploading layer parts. If you do not specify a registry, the default public registry is assumed.

uploadLayerPart_repositoryName :: Lens' UploadLayerPart Text

The name of the repository to which you are uploading layer parts.

uploadLayerPart_uploadId :: Lens' UploadLayerPart Text

The upload ID from a previous InitiateLayerUpload operation to associate with the layer part upload.

uploadLayerPart_partFirstByte :: Lens' UploadLayerPart Natural

The position of the first byte of the layer part witin the overall image layer.

uploadLayerPart_partLastByte :: Lens' UploadLayerPart Natural

The position of the last byte of the layer part within the overall image layer.

uploadLayerPart_layerPartBlob :: Lens' UploadLayerPart ByteString

The base64-encoded layer part payload.-- -- Note: This Lens automatically encodes and decodes Base64 data. -- The underlying isomorphism will encode to Base64 representation during -- serialisation, and decode from Base64 representation during deserialisation. -- This Lens accepts and returns only raw unencoded data.

uploadLayerPartResponse_registryId :: Lens' UploadLayerPartResponse (Maybe Text)

The registry ID associated with the request.

uploadLayerPartResponse_lastByteReceived :: Lens' UploadLayerPartResponse (Maybe Natural)

The integer value of the last byte received in the request.

uploadLayerPartResponse_repositoryName :: Lens' UploadLayerPartResponse (Maybe Text)

The repository name associated with the request.

uploadLayerPartResponse_uploadId :: Lens' UploadLayerPartResponse (Maybe Text)

The upload ID associated with the request.

uploadLayerPartResponse_httpStatus :: Lens' UploadLayerPartResponse Int

The response's http status code.

GetRepositoryCatalogData

getRepositoryCatalogData_registryId :: Lens' GetRepositoryCatalogData (Maybe Text)

The AWS account ID associated with the registry that contains the repositories to be described. If you do not specify a registry, the default public registry is assumed.

getRepositoryCatalogData_repositoryName :: Lens' GetRepositoryCatalogData Text

The name of the repository to retrieve the catalog metadata for.

getRepositoryCatalogDataResponse_catalogData :: Lens' GetRepositoryCatalogDataResponse (Maybe RepositoryCatalogData)

The catalog metadata for the repository.

getRepositoryCatalogDataResponse_httpStatus :: Lens' GetRepositoryCatalogDataResponse Int

The response's http status code.

GetRegistryCatalogData

getRegistryCatalogDataResponse_httpStatus :: Lens' GetRegistryCatalogDataResponse Int

The response's http status code.

getRegistryCatalogDataResponse_registryCatalogData :: Lens' GetRegistryCatalogDataResponse RegistryCatalogData

The catalog metadata for the public registry.

TagResource

tagResource_resourceArn :: Lens' TagResource Text

The Amazon Resource Name (ARN) of the resource to which to add tags. Currently, the supported resource is an Amazon ECR Public repository.

tagResource_tags :: Lens' TagResource [Tag]

The tags to add to the resource. A tag is an array of key-value pairs. Tag keys can have a maximum character length of 128 characters, and tag values can have a maximum length of 256 characters.

tagResourceResponse_httpStatus :: Lens' TagResourceResponse Int

The response's http status code.

SetRepositoryPolicy

setRepositoryPolicy_force :: Lens' SetRepositoryPolicy (Maybe Bool)

If the policy you are attempting to set on a repository policy would prevent you from setting another policy in the future, you must force the SetRepositoryPolicy operation. This is intended to prevent accidental repository lock outs.

setRepositoryPolicy_registryId :: Lens' SetRepositoryPolicy (Maybe Text)

The AWS account ID associated with the registry that contains the repository. If you do not specify a registry, the default public registry is assumed.

setRepositoryPolicy_repositoryName :: Lens' SetRepositoryPolicy Text

The name of the repository to receive the policy.

setRepositoryPolicy_policyText :: Lens' SetRepositoryPolicy Text

The JSON repository policy text to apply to the repository. For more information, see Amazon ECR Repository Policies in the Amazon Elastic Container Registry User Guide.

setRepositoryPolicyResponse_registryId :: Lens' SetRepositoryPolicyResponse (Maybe Text)

The registry ID associated with the request.

setRepositoryPolicyResponse_repositoryName :: Lens' SetRepositoryPolicyResponse (Maybe Text)

The repository name associated with the request.

setRepositoryPolicyResponse_policyText :: Lens' SetRepositoryPolicyResponse (Maybe Text)

The JSON repository policy text applied to the repository.

setRepositoryPolicyResponse_httpStatus :: Lens' SetRepositoryPolicyResponse Int

The response's http status code.

InitiateLayerUpload

initiateLayerUpload_registryId :: Lens' InitiateLayerUpload (Maybe Text)

The AWS account ID associated with the registry to which you intend to upload layers. If you do not specify a registry, the default public registry is assumed.

initiateLayerUpload_repositoryName :: Lens' InitiateLayerUpload Text

The name of the repository to which you intend to upload layers.

initiateLayerUploadResponse_partSize :: Lens' InitiateLayerUploadResponse (Maybe Natural)

The size, in bytes, that Amazon ECR expects future layer part uploads to be.

initiateLayerUploadResponse_uploadId :: Lens' InitiateLayerUploadResponse (Maybe Text)

The upload ID for the layer upload. This parameter is passed to further UploadLayerPart and CompleteLayerUpload operations.

initiateLayerUploadResponse_httpStatus :: Lens' InitiateLayerUploadResponse Int

The response's http status code.

UntagResource

untagResource_resourceArn :: Lens' UntagResource Text

The Amazon Resource Name (ARN) of the resource from which to delete tags. Currently, the supported resource is an Amazon ECR Public repository.

untagResource_tagKeys :: Lens' UntagResource [Text]

The keys of the tags to be removed.

untagResourceResponse_httpStatus :: Lens' UntagResourceResponse Int

The response's http status code.

DescribeImageTags

describeImageTags_registryId :: Lens' DescribeImageTags (Maybe Text)

The AWS account ID associated with the public registry that contains the repository in which to describe images. If you do not specify a registry, the default public registry is assumed.

describeImageTags_nextToken :: Lens' DescribeImageTags (Maybe Text)

The nextToken value returned from a previous paginated DescribeImageTags request where maxResults was used and the results exceeded the value of that parameter. Pagination continues from the end of the previous results that returned the nextToken value. This value is null when there are no more results to return. This option cannot be used when you specify images with imageIds.

describeImageTags_maxResults :: Lens' DescribeImageTags (Maybe Natural)

The maximum number of repository results returned by DescribeImageTags in paginated output. When this parameter is used, DescribeImageTags only returns maxResults results in a single page along with a nextToken response element. The remaining results of the initial request can be seen by sending another DescribeImageTags request with the returned nextToken value. This value can be between 1 and 1000. If this parameter is not used, then DescribeImageTags returns up to 100 results and a nextToken value, if applicable. This option cannot be used when you specify images with imageIds.

describeImageTags_repositoryName :: Lens' DescribeImageTags Text

The name of the repository that contains the image tag details to describe.

describeImageTagsResponse_nextToken :: Lens' DescribeImageTagsResponse (Maybe Text)

The nextToken value to include in a future DescribeImageTags request. When the results of a DescribeImageTags request exceed maxResults, this value can be used to retrieve the next page of results. This value is null when there are no more results to return.

describeImageTagsResponse_imageTagDetails :: Lens' DescribeImageTagsResponse (Maybe [ImageTagDetail])

The image tag details for the images in the requested repository.

describeImageTagsResponse_httpStatus :: Lens' DescribeImageTagsResponse Int

The response's http status code.

DeleteRepository

deleteRepository_force :: Lens' DeleteRepository (Maybe Bool)

If a repository contains images, forces the deletion.

deleteRepository_registryId :: Lens' DeleteRepository (Maybe Text)

The AWS account ID associated with the public registry that contains the repository to delete. If you do not specify a registry, the default public registry is assumed.

deleteRepository_repositoryName :: Lens' DeleteRepository Text

The name of the repository to delete.

deleteRepositoryResponse_repository :: Lens' DeleteRepositoryResponse (Maybe Repository)

The repository that was deleted.

deleteRepositoryResponse_httpStatus :: Lens' DeleteRepositoryResponse Int

The response's http status code.

PutImage

putImage_registryId :: Lens' PutImage (Maybe Text)

The AWS account ID associated with the public registry that contains the repository in which to put the image. If you do not specify a registry, the default public registry is assumed.

putImage_imageManifestMediaType :: Lens' PutImage (Maybe Text)

The media type of the image manifest. If you push an image manifest that does not contain the mediaType field, you must specify the imageManifestMediaType in the request.

putImage_imageDigest :: Lens' PutImage (Maybe Text)

The image digest of the image manifest corresponding to the image.

putImage_imageTag :: Lens' PutImage (Maybe Text)

The tag to associate with the image. This parameter is required for images that use the Docker Image Manifest V2 Schema 2 or Open Container Initiative (OCI) formats.

putImage_repositoryName :: Lens' PutImage Text

The name of the repository in which to put the image.

putImage_imageManifest :: Lens' PutImage Text

The image manifest corresponding to the image to be uploaded.

putImageResponse_image :: Lens' PutImageResponse (Maybe Image)

Details of the image uploaded.

putImageResponse_httpStatus :: Lens' PutImageResponse Int

The response's http status code.

GetAuthorizationToken

getAuthorizationTokenResponse_authorizationData :: Lens' GetAuthorizationTokenResponse (Maybe AuthorizationData)

An authorization token data object that corresponds to a public registry.

getAuthorizationTokenResponse_httpStatus :: Lens' GetAuthorizationTokenResponse Int

The response's http status code.

DescribeImages

describeImages_registryId :: Lens' DescribeImages (Maybe Text)

The AWS account ID associated with the public registry that contains the repository in which to describe images. If you do not specify a registry, the default public registry is assumed.

describeImages_imageIds :: Lens' DescribeImages (Maybe (NonEmpty ImageIdentifier))

The list of image IDs for the requested repository.

describeImages_nextToken :: Lens' DescribeImages (Maybe Text)

The nextToken value returned from a previous paginated DescribeImages request where maxResults was used and the results exceeded the value of that parameter. Pagination continues from the end of the previous results that returned the nextToken value. This value is null when there are no more results to return. This option cannot be used when you specify images with imageIds.

describeImages_maxResults :: Lens' DescribeImages (Maybe Natural)

The maximum number of repository results returned by DescribeImages in paginated output. When this parameter is used, DescribeImages only returns maxResults results in a single page along with a nextToken response element. The remaining results of the initial request can be seen by sending another DescribeImages request with the returned nextToken value. This value can be between 1 and 1000. If this parameter is not used, then DescribeImages returns up to 100 results and a nextToken value, if applicable. This option cannot be used when you specify images with imageIds.

describeImages_repositoryName :: Lens' DescribeImages Text

The repository that contains the images to describe.

describeImagesResponse_imageDetails :: Lens' DescribeImagesResponse (Maybe [ImageDetail])

A list of ImageDetail objects that contain data about the image.

describeImagesResponse_nextToken :: Lens' DescribeImagesResponse (Maybe Text)

The nextToken value to include in a future DescribeImages request. When the results of a DescribeImages request exceed maxResults, this value can be used to retrieve the next page of results. This value is null when there are no more results to return.

describeImagesResponse_httpStatus :: Lens' DescribeImagesResponse Int

The response's http status code.

Types

AuthorizationData

authorizationData_expiresAt :: Lens' AuthorizationData (Maybe UTCTime)

The Unix time in seconds and milliseconds when the authorization token expires. Authorization tokens are valid for 12 hours.

authorizationData_authorizationToken :: Lens' AuthorizationData (Maybe Text)

A base64-encoded string that contains authorization data for a public Amazon ECR registry. When the string is decoded, it is presented in the format user:password for public registry authentication using docker login.

Image

image_registryId :: Lens' Image (Maybe Text)

The AWS account ID associated with the registry containing the image.

image_imageManifestMediaType :: Lens' Image (Maybe Text)

The manifest media type of the image.

image_imageId :: Lens' Image (Maybe ImageIdentifier)

An object containing the image tag and image digest associated with an image.

image_repositoryName :: Lens' Image (Maybe Text)

The name of the repository associated with the image.

image_imageManifest :: Lens' Image (Maybe Text)

The image manifest associated with the image.

ImageDetail

imageDetail_registryId :: Lens' ImageDetail (Maybe Text)

The AWS account ID associated with the public registry to which this image belongs.

imageDetail_imageTags :: Lens' ImageDetail (Maybe [Text])

The list of tags associated with this image.

imageDetail_imageManifestMediaType :: Lens' ImageDetail (Maybe Text)

The media type of the image manifest.

imageDetail_imageSizeInBytes :: Lens' ImageDetail (Maybe Integer)

The size, in bytes, of the image in the repository.

If the image is a manifest list, this will be the max size of all manifests in the list.

Beginning with Docker version 1.9, the Docker client compresses image layers before pushing them to a V2 Docker registry. The output of the docker images command shows the uncompressed image size, so it may return a larger image size than the image sizes returned by DescribeImages.

imageDetail_imageDigest :: Lens' ImageDetail (Maybe Text)

The sha256 digest of the image manifest.

imageDetail_artifactMediaType :: Lens' ImageDetail (Maybe Text)

The artifact media type of the image.

imageDetail_imagePushedAt :: Lens' ImageDetail (Maybe UTCTime)

The date and time, expressed in standard JavaScript date format, at which the current image was pushed to the repository.

imageDetail_repositoryName :: Lens' ImageDetail (Maybe Text)

The name of the repository to which this image belongs.

ImageFailure

imageFailure_failureReason :: Lens' ImageFailure (Maybe Text)

The reason for the failure.

imageFailure_failureCode :: Lens' ImageFailure (Maybe ImageFailureCode)

The code associated with the failure.

imageFailure_imageId :: Lens' ImageFailure (Maybe ImageIdentifier)

The image ID associated with the failure.

ImageIdentifier

imageIdentifier_imageDigest :: Lens' ImageIdentifier (Maybe Text)

The sha256 digest of the image manifest.

imageIdentifier_imageTag :: Lens' ImageIdentifier (Maybe Text)

The tag used for the image.

ImageTagDetail

imageTagDetail_createdAt :: Lens' ImageTagDetail (Maybe UTCTime)

The time stamp indicating when the image tag was created.

imageTagDetail_imageDetail :: Lens' ImageTagDetail (Maybe ReferencedImageDetail)

An object that describes the details of an image.

imageTagDetail_imageTag :: Lens' ImageTagDetail (Maybe Text)

The tag associated with the image.

Layer

layer_mediaType :: Lens' Layer (Maybe Text)

The media type of the layer, such as application/vnd.docker.image.rootfs.diff.tar.gzip or application/vnd.oci.image.layer.v1.tar+gzip.

layer_layerDigest :: Lens' Layer (Maybe Text)

The sha256 digest of the image layer.

layer_layerSize :: Lens' Layer (Maybe Integer)

The size, in bytes, of the image layer.

layer_layerAvailability :: Lens' Layer (Maybe LayerAvailability)

The availability status of the image layer.

LayerFailure

layerFailure_failureReason :: Lens' LayerFailure (Maybe Text)

The reason for the failure.

layerFailure_failureCode :: Lens' LayerFailure (Maybe LayerFailureCode)

The failure code associated with the failure.

layerFailure_layerDigest :: Lens' LayerFailure (Maybe Text)

The layer digest associated with the failure.

ReferencedImageDetail

referencedImageDetail_imageManifestMediaType :: Lens' ReferencedImageDetail (Maybe Text)

The media type of the image manifest.

referencedImageDetail_imageSizeInBytes :: Lens' ReferencedImageDetail (Maybe Integer)

The size, in bytes, of the image in the repository.

If the image is a manifest list, this will be the max size of all manifests in the list.

Beginning with Docker version 1.9, the Docker client compresses image layers before pushing them to a V2 Docker registry. The output of the docker images command shows the uncompressed image size, so it may return a larger image size than the image sizes returned by DescribeImages.

referencedImageDetail_imageDigest :: Lens' ReferencedImageDetail (Maybe Text)

The sha256 digest of the image manifest.

referencedImageDetail_artifactMediaType :: Lens' ReferencedImageDetail (Maybe Text)

The artifact media type of the image.

referencedImageDetail_imagePushedAt :: Lens' ReferencedImageDetail (Maybe UTCTime)

The date and time, expressed in standard JavaScript date format, at which the current image tag was pushed to the repository.

Registry

registry_registryId :: Lens' Registry Text

The AWS account ID associated with the registry. If you do not specify a registry, the default public registry is assumed.

registry_registryArn :: Lens' Registry Text

The Amazon Resource Name (ARN) of the public registry.

registry_registryUri :: Lens' Registry Text

The URI of a public registry. The URI contains a universal prefix and the registry alias.

registry_verified :: Lens' Registry Bool

Whether the account is verified. This indicates whether the account is an AWS Marketplace vendor. If an account is verified, each public repository will received a verified account badge on the Amazon ECR Public Gallery.

registry_aliases :: Lens' Registry [RegistryAlias]

An array of objects representing the aliases for a public registry.

RegistryAlias

registryAlias_name :: Lens' RegistryAlias Text

The name of the registry alias.

registryAlias_status :: Lens' RegistryAlias RegistryAliasStatus

The status of the registry alias.

registryAlias_primaryRegistryAlias :: Lens' RegistryAlias Bool

Whether or not the registry alias is the primary alias for the registry. If true, the alias is the primary registry alias and is displayed in both the repository URL and the image URI used in the docker pull commands on the Amazon ECR Public Gallery.

A registry alias that is not the primary registry alias can be used in the repository URI in a docker pull command.

registryAlias_defaultRegistryAlias :: Lens' RegistryAlias Bool

Whether or not the registry alias is the default alias for the registry. When the first public repository is created, your public registry is assigned a default registry alias.

RegistryCatalogData

registryCatalogData_displayName :: Lens' RegistryCatalogData (Maybe Text)

The display name for a public registry. This appears on the Amazon ECR Public Gallery.

Only accounts that have the verified account badge can have a registry display name.

Repository

repository_repositoryArn :: Lens' Repository (Maybe Text)

The Amazon Resource Name (ARN) that identifies the repository. The ARN contains the arn:aws:ecr namespace, followed by the region of the repository, AWS account ID of the repository owner, repository namespace, and repository name. For example, arn:aws:ecr:region:012345678910:repository/test.

repository_createdAt :: Lens' Repository (Maybe UTCTime)

The date and time, in JavaScript date format, when the repository was created.

repository_registryId :: Lens' Repository (Maybe Text)

The AWS account ID associated with the public registry that contains the repository.

repository_repositoryUri :: Lens' Repository (Maybe Text)

The URI for the repository. You can use this URI for container image push and pull operations.

repository_repositoryName :: Lens' Repository (Maybe Text)

The name of the repository.

RepositoryCatalogData

repositoryCatalogData_logoUrl :: Lens' RepositoryCatalogData (Maybe Text)

The URL containing the logo associated with the repository.

repositoryCatalogData_architectures :: Lens' RepositoryCatalogData (Maybe [Text])

The architecture tags that are associated with the repository.

Only supported operating system tags appear publicly in the Amazon ECR Public Gallery. For more information, see RepositoryCatalogDataInput.

repositoryCatalogData_usageText :: Lens' RepositoryCatalogData (Maybe Text)

The longform usage details of the contents of the repository. The usage text provides context for users of the repository.

repositoryCatalogData_marketplaceCertified :: Lens' RepositoryCatalogData (Maybe Bool)

Whether or not the repository is certified by AWS Marketplace.

repositoryCatalogData_aboutText :: Lens' RepositoryCatalogData (Maybe Text)

The longform description of the contents of the repository. This text appears in the repository details on the Amazon ECR Public Gallery.

repositoryCatalogData_operatingSystems :: Lens' RepositoryCatalogData (Maybe [Text])

The operating system tags that are associated with the repository.

Only supported operating system tags appear publicly in the Amazon ECR Public Gallery. For more information, see RepositoryCatalogDataInput.

repositoryCatalogData_description :: Lens' RepositoryCatalogData (Maybe Text)

The short description of the repository.

RepositoryCatalogDataInput

repositoryCatalogDataInput_logoImageBlob :: Lens' RepositoryCatalogDataInput (Maybe ByteString)

The base64-encoded repository logo payload.

The repository logo is only publicly visible in the Amazon ECR Public Gallery for verified accounts.-- -- Note: This Lens automatically encodes and decodes Base64 data. -- The underlying isomorphism will encode to Base64 representation during -- serialisation, and decode from Base64 representation during deserialisation. -- This Lens accepts and returns only raw unencoded data.

repositoryCatalogDataInput_architectures :: Lens' RepositoryCatalogDataInput (Maybe [Text])

The system architecture that the images in the repository are compatible with. On the Amazon ECR Public Gallery, the following supported architectures will appear as badges on the repository and are used as search filters.

Linux

Windows

If an unsupported tag is added to your repository catalog data, it will be associated with the repository and can be retrieved using the API but will not be discoverable in the Amazon ECR Public Gallery.

repositoryCatalogDataInput_usageText :: Lens' RepositoryCatalogDataInput (Maybe Text)

Detailed information on how to use the contents of the repository. It is publicly visible in the Amazon ECR Public Gallery. The usage text provides context, support information, and additional usage details for users of the repository. The text must be in markdown format.

repositoryCatalogDataInput_aboutText :: Lens' RepositoryCatalogDataInput (Maybe Text)

A detailed description of the contents of the repository. It is publicly visible in the Amazon ECR Public Gallery. The text must be in markdown format.

repositoryCatalogDataInput_operatingSystems :: Lens' RepositoryCatalogDataInput (Maybe [Text])

The operating systems that the images in the repository are compatible with. On the Amazon ECR Public Gallery, the following supported operating systems will appear as badges on the repository and are used as search filters.

ARM

ARM 64

x86

x86-64

If an unsupported tag is added to your repository catalog data, it will be associated with the repository and can be retrieved using the API but will not be discoverable in the Amazon ECR Public Gallery.

repositoryCatalogDataInput_description :: Lens' RepositoryCatalogDataInput (Maybe Text)

A short description of the contents of the repository. This text appears in both the image details and also when searching for repositories on the Amazon ECR Public Gallery.

Tag

tag_value :: Lens' Tag (Maybe Text)

The optional part of a key-value pair that make up a tag. A value acts as a descriptor within a tag category (key).

tag_key :: Lens' Tag (Maybe Text)

One part of a key-value pair that make up a tag. A key is a general label that acts like a category for more specific tag values.