Amazonka.ECRPublic.PutRegistryCatalogData

Description

Create or updates the catalog data for a public registry.

Creating a Request

data PutRegistryCatalogData

See: newPutRegistryCatalogData smart constructor.

Constructors

PutRegistryCatalogData'

Fields

displayName :: Maybe Text

The display name for a public registry. The display name is shown as the repository author in the Amazon ECR Public Gallery.

The registry display name is only publicly visible in the Amazon ECR Public Gallery for verified accounts.

newPutRegistryCatalogData :: PutRegistryCatalogData

Create a value of PutRegistryCatalogData with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:displayName:PutRegistryCatalogData', putRegistryCatalogData_displayName - The display name for a public registry. The display name is shown as the repository author in the Amazon ECR Public Gallery.

The registry display name is only publicly visible in the Amazon ECR Public Gallery for verified accounts.

Request Lenses

putRegistryCatalogData_displayName :: Lens' PutRegistryCatalogData (Maybe Text)

The display name for a public registry. The display name is shown as the repository author in the Amazon ECR Public Gallery.

The registry display name is only publicly visible in the Amazon ECR Public Gallery for verified accounts.

Destructuring the Response

data PutRegistryCatalogDataResponse

See: newPutRegistryCatalogDataResponse smart constructor.

Constructors

PutRegistryCatalogDataResponse'

Fields

httpStatus :: Int

The response's http status code.

registryCatalogData :: RegistryCatalogData

The catalog data for the public registry.

newPutRegistryCatalogDataResponse

Arguments

:: Int
$sel:httpStatus:PutRegistryCatalogDataResponse'

-> RegistryCatalogData
$sel:registryCatalogData:PutRegistryCatalogDataResponse'

-> PutRegistryCatalogDataResponse

Create a value of PutRegistryCatalogDataResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:httpStatus:PutRegistryCatalogDataResponse', putRegistryCatalogDataResponse_httpStatus - The response's http status code.

$sel:registryCatalogData:PutRegistryCatalogDataResponse', putRegistryCatalogDataResponse_registryCatalogData - The catalog data for the public registry.

Response Lenses

putRegistryCatalogDataResponse_httpStatus :: Lens' PutRegistryCatalogDataResponse Int

The response's http status code.

putRegistryCatalogDataResponse_registryCatalogData :: Lens' PutRegistryCatalogDataResponse RegistryCatalogData

The catalog data for the public registry.