Amazonka.ECRPublic.Types.Image

Description

Documentation

data Image

An object representing an Amazon ECR image.

See: newImage smart constructor.

Constructors

Image'

Fields

registryId :: Maybe Text

The AWS account ID associated with the registry containing the image.

imageManifestMediaType :: Maybe Text

The manifest media type of the image.

imageId :: Maybe ImageIdentifier

An object containing the image tag and image digest associated with an image.

repositoryName :: Maybe Text

The name of the repository associated with the image.

imageManifest :: Maybe Text

The image manifest associated with the image.

newImage :: Image

Create a value of Image with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:registryId:Image', image_registryId - The AWS account ID associated with the registry containing the image.

$sel:imageManifestMediaType:Image', image_imageManifestMediaType - The manifest media type of the image.

$sel:imageId:Image', image_imageId - An object containing the image tag and image digest associated with an image.

$sel:repositoryName:Image', image_repositoryName - The name of the repository associated with the image.

$sel:imageManifest:Image', image_imageManifest - The image manifest associated with the image.

image_registryId :: Lens' Image (Maybe Text)

The AWS account ID associated with the registry containing the image.

image_imageManifestMediaType :: Lens' Image (Maybe Text)

The manifest media type of the image.

image_imageId :: Lens' Image (Maybe ImageIdentifier)

An object containing the image tag and image digest associated with an image.

image_repositoryName :: Lens' Image (Maybe Text)

The name of the repository associated with the image.

image_imageManifest :: Lens' Image (Maybe Text)

The image manifest associated with the image.