Amazonka.ECRPublic.Types.Layer

Description

Documentation

data Layer

An object representing an Amazon ECR image layer.

See: newLayer smart constructor.

Constructors

Layer'

Fields

mediaType :: Maybe Text

The media type of the layer, such as application/vnd.docker.image.rootfs.diff.tar.gzip or application/vnd.oci.image.layer.v1.tar+gzip.

layerDigest :: Maybe Text

The sha256 digest of the image layer.

layerSize :: Maybe Integer

The size, in bytes, of the image layer.

layerAvailability :: Maybe LayerAvailability

The availability status of the image layer.

newLayer :: Layer

Create a value of Layer with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:mediaType:Layer', layer_mediaType - The media type of the layer, such as application/vnd.docker.image.rootfs.diff.tar.gzip or application/vnd.oci.image.layer.v1.tar+gzip.

$sel:layerDigest:Layer', layer_layerDigest - The sha256 digest of the image layer.

$sel:layerSize:Layer', layer_layerSize - The size, in bytes, of the image layer.

$sel:layerAvailability:Layer', layer_layerAvailability - The availability status of the image layer.

layer_mediaType :: Lens' Layer (Maybe Text)

The media type of the layer, such as application/vnd.docker.image.rootfs.diff.tar.gzip or application/vnd.oci.image.layer.v1.tar+gzip.

layer_layerDigest :: Lens' Layer (Maybe Text)

The sha256 digest of the image layer.

layer_layerSize :: Lens' Layer (Maybe Integer)

The size, in bytes, of the image layer.

layer_layerAvailability :: Lens' Layer (Maybe LayerAvailability)

The availability status of the image layer.