Amazonka.ECRPublic.Types.RepositoryCatalogData

Description

Documentation

data RepositoryCatalogData

The catalog data for a repository. This data is publicly visible in the Amazon ECR Public Gallery.

See: newRepositoryCatalogData smart constructor.

Constructors

RepositoryCatalogData'

Fields

logoUrl :: Maybe Text

The URL containing the logo associated with the repository.

architectures :: Maybe [Text]

The architecture tags that are associated with the repository.

Only supported operating system tags appear publicly in the Amazon ECR Public Gallery. For more information, see RepositoryCatalogDataInput.

usageText :: Maybe Text

The longform usage details of the contents of the repository. The usage text provides context for users of the repository.

marketplaceCertified :: Maybe Bool

Whether or not the repository is certified by AWS Marketplace.

aboutText :: Maybe Text

The longform description of the contents of the repository. This text appears in the repository details on the Amazon ECR Public Gallery.

operatingSystems :: Maybe [Text]

The operating system tags that are associated with the repository.

Only supported operating system tags appear publicly in the Amazon ECR Public Gallery. For more information, see RepositoryCatalogDataInput.

description :: Maybe Text

The short description of the repository.

newRepositoryCatalogData :: RepositoryCatalogData

Create a value of RepositoryCatalogData with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:logoUrl:RepositoryCatalogData', repositoryCatalogData_logoUrl - The URL containing the logo associated with the repository.

$sel:architectures:RepositoryCatalogData', repositoryCatalogData_architectures - The architecture tags that are associated with the repository.

Only supported operating system tags appear publicly in the Amazon ECR Public Gallery. For more information, see RepositoryCatalogDataInput.

$sel:usageText:RepositoryCatalogData', repositoryCatalogData_usageText - The longform usage details of the contents of the repository. The usage text provides context for users of the repository.

$sel:marketplaceCertified:RepositoryCatalogData', repositoryCatalogData_marketplaceCertified - Whether or not the repository is certified by AWS Marketplace.

$sel:aboutText:RepositoryCatalogData', repositoryCatalogData_aboutText - The longform description of the contents of the repository. This text appears in the repository details on the Amazon ECR Public Gallery.

$sel:operatingSystems:RepositoryCatalogData', repositoryCatalogData_operatingSystems - The operating system tags that are associated with the repository.

Only supported operating system tags appear publicly in the Amazon ECR Public Gallery. For more information, see RepositoryCatalogDataInput.

$sel:description:RepositoryCatalogData', repositoryCatalogData_description - The short description of the repository.

repositoryCatalogData_logoUrl :: Lens' RepositoryCatalogData (Maybe Text)

The URL containing the logo associated with the repository.

repositoryCatalogData_architectures :: Lens' RepositoryCatalogData (Maybe [Text])

The architecture tags that are associated with the repository.

Only supported operating system tags appear publicly in the Amazon ECR Public Gallery. For more information, see RepositoryCatalogDataInput.

repositoryCatalogData_usageText :: Lens' RepositoryCatalogData (Maybe Text)

The longform usage details of the contents of the repository. The usage text provides context for users of the repository.

repositoryCatalogData_marketplaceCertified :: Lens' RepositoryCatalogData (Maybe Bool)

Whether or not the repository is certified by AWS Marketplace.

repositoryCatalogData_aboutText :: Lens' RepositoryCatalogData (Maybe Text)

The longform description of the contents of the repository. This text appears in the repository details on the Amazon ECR Public Gallery.

repositoryCatalogData_operatingSystems :: Lens' RepositoryCatalogData (Maybe [Text])

The operating system tags that are associated with the repository.

Only supported operating system tags appear publicly in the Amazon ECR Public Gallery. For more information, see RepositoryCatalogDataInput.

repositoryCatalogData_description :: Lens' RepositoryCatalogData (Maybe Text)

The short description of the repository.