Amazonka.ECS.DeregisterTaskDefinition

Description

Deregisters the specified task definition by family and revision. Upon deregistration, the task definition is marked as INACTIVE. Existing tasks and services that reference an INACTIVE task definition continue to run without disruption. Existing services that reference an INACTIVE task definition can still scale up or down by modifying the service's desired count.

You cannot use an INACTIVE task definition to run new tasks or create new services, and you cannot update an existing service to reference an INACTIVE task definition. However, there may be up to a 10-minute window following deregistration where these restrictions have not yet taken effect.

At this time, INACTIVE task definitions remain discoverable in your account indefinitely. However, this behavior is subject to change in the future, so you should not rely on INACTIVE task definitions persisting beyond the lifecycle of any associated tasks and services.

Creating a Request

data DeregisterTaskDefinition

See: newDeregisterTaskDefinition smart constructor.

Constructors

DeregisterTaskDefinition'

Fields

taskDefinition :: Text

The family and revision (family:revision) or full Amazon Resource Name (ARN) of the task definition to deregister. You must specify a revision.

newDeregisterTaskDefinition

Arguments

:: Text
$sel:taskDefinition:DeregisterTaskDefinition'

-> DeregisterTaskDefinition

Create a value of DeregisterTaskDefinition with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:taskDefinition:DeregisterTaskDefinition', deregisterTaskDefinition_taskDefinition - The family and revision (family:revision) or full Amazon Resource Name (ARN) of the task definition to deregister. You must specify a revision.

Request Lenses

deregisterTaskDefinition_taskDefinition :: Lens' DeregisterTaskDefinition Text

The family and revision (family:revision) or full Amazon Resource Name (ARN) of the task definition to deregister. You must specify a revision.

Destructuring the Response

data DeregisterTaskDefinitionResponse

See: newDeregisterTaskDefinitionResponse smart constructor.

Constructors

DeregisterTaskDefinitionResponse'

Fields

taskDefinition :: Maybe TaskDefinition

The full description of the deregistered task.

httpStatus :: Int

The response's http status code.

newDeregisterTaskDefinitionResponse

Arguments

:: Int
$sel:httpStatus:DeregisterTaskDefinitionResponse'

-> DeregisterTaskDefinitionResponse

Create a value of DeregisterTaskDefinitionResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:taskDefinition:DeregisterTaskDefinition', deregisterTaskDefinitionResponse_taskDefinition - The full description of the deregistered task.

$sel:httpStatus:DeregisterTaskDefinitionResponse', deregisterTaskDefinitionResponse_httpStatus - The response's http status code.

Response Lenses

deregisterTaskDefinitionResponse_taskDefinition :: Lens' DeregisterTaskDefinitionResponse (Maybe TaskDefinition)

The full description of the deregistered task.

deregisterTaskDefinitionResponse_httpStatus :: Lens' DeregisterTaskDefinitionResponse Int

The response's http status code.