Amazonka.ECS.DescribeServices

Description

Describes the specified services running in your cluster.

Creating a Request

data DescribeServices

See: newDescribeServices smart constructor.

Constructors

DescribeServices'

Fields

include :: Maybe [ServiceField]

Specifies whether you want to see the resource tags for the service. If TAGS is specified, the tags are included in the response. If this field is omitted, tags are not included in the response.

cluster :: Maybe Text

The short name or full Amazon Resource Name (ARN)the cluster that hosts the service to describe. If you do not specify a cluster, the default cluster is assumed. This parameter is required if the service or services you are describing were launched in any cluster other than the default cluster.

services :: [Text]

A list of services to describe. You may specify up to 10 services to describe in a single operation.

newDescribeServices :: DescribeServices

Create a value of DescribeServices with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:include:DescribeServices', describeServices_include - Specifies whether you want to see the resource tags for the service. If TAGS is specified, the tags are included in the response. If this field is omitted, tags are not included in the response.

$sel:cluster:DescribeServices', describeServices_cluster - The short name or full Amazon Resource Name (ARN)the cluster that hosts the service to describe. If you do not specify a cluster, the default cluster is assumed. This parameter is required if the service or services you are describing were launched in any cluster other than the default cluster.

$sel:services:DescribeServices', describeServices_services - A list of services to describe. You may specify up to 10 services to describe in a single operation.

Request Lenses

describeServices_include :: Lens' DescribeServices (Maybe [ServiceField])

Specifies whether you want to see the resource tags for the service. If TAGS is specified, the tags are included in the response. If this field is omitted, tags are not included in the response.

describeServices_cluster :: Lens' DescribeServices (Maybe Text)

The short name or full Amazon Resource Name (ARN)the cluster that hosts the service to describe. If you do not specify a cluster, the default cluster is assumed. This parameter is required if the service or services you are describing were launched in any cluster other than the default cluster.

describeServices_services :: Lens' DescribeServices [Text]

A list of services to describe. You may specify up to 10 services to describe in a single operation.

Destructuring the Response

data DescribeServicesResponse

See: newDescribeServicesResponse smart constructor.

Constructors

DescribeServicesResponse'

Fields

failures :: Maybe [Failure]

Any failures associated with the call.

services :: Maybe [ContainerService]

The list of services described.

httpStatus :: Int

The response's http status code.

newDescribeServicesResponse

Arguments

:: Int
$sel:httpStatus:DescribeServicesResponse'

-> DescribeServicesResponse

Create a value of DescribeServicesResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:failures:DescribeServicesResponse', describeServicesResponse_failures - Any failures associated with the call.

$sel:services:DescribeServices', describeServicesResponse_services - The list of services described.

$sel:httpStatus:DescribeServicesResponse', describeServicesResponse_httpStatus - The response's http status code.

Response Lenses

describeServicesResponse_failures :: Lens' DescribeServicesResponse (Maybe [Failure])

Any failures associated with the call.

describeServicesResponse_services :: Lens' DescribeServicesResponse (Maybe [ContainerService])

The list of services described.

describeServicesResponse_httpStatus :: Lens' DescribeServicesResponse Int

The response's http status code.