Amazonka.ECS.SubmitTaskStateChange

Description

This action is only used by the Amazon ECS agent, and it is not intended for use outside of the agent.

Sent to acknowledge that a task changed states.

Creating a Request

data SubmitTaskStateChange

See: newSubmitTaskStateChange smart constructor.

Constructors

SubmitTaskStateChange'

Fields

status :: Maybe Text

The status of the state change request.

managedAgents :: Maybe [ManagedAgentStateChange]

The details for the managed agent associated with the task.

cluster :: Maybe Text

The short name or full Amazon Resource Name (ARN) of the cluster that hosts the task.

attachments :: Maybe [AttachmentStateChange]

Any attachments associated with the state change request.

executionStoppedAt :: Maybe POSIX

The Unix timestamp for when the task execution stopped.

pullStoppedAt :: Maybe POSIX

The Unix timestamp for when the container image pull completed.

containers :: Maybe [ContainerStateChange]

Any containers associated with the state change request.

reason :: Maybe Text

The reason for the state change request.

task :: Maybe Text

The task ID or full ARN of the task in the state change request.

pullStartedAt :: Maybe POSIX

The Unix timestamp for when the container image pull began.

newSubmitTaskStateChange :: SubmitTaskStateChange

Create a value of SubmitTaskStateChange with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:status:SubmitTaskStateChange', submitTaskStateChange_status - The status of the state change request.

$sel:managedAgents:SubmitTaskStateChange', submitTaskStateChange_managedAgents - The details for the managed agent associated with the task.

$sel:cluster:SubmitTaskStateChange', submitTaskStateChange_cluster - The short name or full Amazon Resource Name (ARN) of the cluster that hosts the task.

$sel:attachments:SubmitTaskStateChange', submitTaskStateChange_attachments - Any attachments associated with the state change request.

$sel:executionStoppedAt:SubmitTaskStateChange', submitTaskStateChange_executionStoppedAt - The Unix timestamp for when the task execution stopped.

$sel:pullStoppedAt:SubmitTaskStateChange', submitTaskStateChange_pullStoppedAt - The Unix timestamp for when the container image pull completed.

$sel:containers:SubmitTaskStateChange', submitTaskStateChange_containers - Any containers associated with the state change request.

$sel:reason:SubmitTaskStateChange', submitTaskStateChange_reason - The reason for the state change request.

$sel:task:SubmitTaskStateChange', submitTaskStateChange_task - The task ID or full ARN of the task in the state change request.

$sel:pullStartedAt:SubmitTaskStateChange', submitTaskStateChange_pullStartedAt - The Unix timestamp for when the container image pull began.

Request Lenses

submitTaskStateChange_status :: Lens' SubmitTaskStateChange (Maybe Text)

The status of the state change request.

submitTaskStateChange_managedAgents :: Lens' SubmitTaskStateChange (Maybe [ManagedAgentStateChange])

The details for the managed agent associated with the task.

submitTaskStateChange_cluster :: Lens' SubmitTaskStateChange (Maybe Text)

The short name or full Amazon Resource Name (ARN) of the cluster that hosts the task.

submitTaskStateChange_attachments :: Lens' SubmitTaskStateChange (Maybe [AttachmentStateChange])

Any attachments associated with the state change request.

submitTaskStateChange_executionStoppedAt :: Lens' SubmitTaskStateChange (Maybe UTCTime)

The Unix timestamp for when the task execution stopped.

submitTaskStateChange_pullStoppedAt :: Lens' SubmitTaskStateChange (Maybe UTCTime)

The Unix timestamp for when the container image pull completed.

submitTaskStateChange_containers :: Lens' SubmitTaskStateChange (Maybe [ContainerStateChange])

Any containers associated with the state change request.

submitTaskStateChange_reason :: Lens' SubmitTaskStateChange (Maybe Text)

The reason for the state change request.

submitTaskStateChange_task :: Lens' SubmitTaskStateChange (Maybe Text)

The task ID or full ARN of the task in the state change request.

submitTaskStateChange_pullStartedAt :: Lens' SubmitTaskStateChange (Maybe UTCTime)

The Unix timestamp for when the container image pull began.

Destructuring the Response

data SubmitTaskStateChangeResponse

See: newSubmitTaskStateChangeResponse smart constructor.

Constructors

SubmitTaskStateChangeResponse'

Fields

acknowledgment :: Maybe Text

Acknowledgement of the state change.

httpStatus :: Int

The response's http status code.

newSubmitTaskStateChangeResponse

Arguments

:: Int
$sel:httpStatus:SubmitTaskStateChangeResponse'

-> SubmitTaskStateChangeResponse

Create a value of SubmitTaskStateChangeResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:acknowledgment:SubmitTaskStateChangeResponse', submitTaskStateChangeResponse_acknowledgment - Acknowledgement of the state change.

$sel:httpStatus:SubmitTaskStateChangeResponse', submitTaskStateChangeResponse_httpStatus - The response's http status code.

Response Lenses

submitTaskStateChangeResponse_acknowledgment :: Lens' SubmitTaskStateChangeResponse (Maybe Text)

Acknowledgement of the state change.

submitTaskStateChangeResponse_httpStatus :: Lens' SubmitTaskStateChangeResponse Int

The response's http status code.