Amazonka.ECS.Types.AutoScalingGroupProvider

Description

Documentation

data AutoScalingGroupProvider

The details of the Auto Scaling group for the capacity provider.

See: newAutoScalingGroupProvider smart constructor.

Constructors

AutoScalingGroupProvider'

Fields

managedScaling :: Maybe ManagedScaling

The managed scaling settings for the Auto Scaling group capacity provider.

managedTerminationProtection :: Maybe ManagedTerminationProtection

The managed termination protection setting to use for the Auto Scaling group capacity provider. This determines whether the Auto Scaling group has managed termination protection.

When using managed termination protection, managed scaling must also be used otherwise managed termination protection will not work.

When managed termination protection is enabled, Amazon ECS prevents the Amazon EC2 instances in an Auto Scaling group that contain tasks from being terminated during a scale-in action. The Auto Scaling group and each instance in the Auto Scaling group must have instance protection from scale-in actions enabled as well. For more information, see Instance Protection in the Auto Scaling User Guide.

When managed termination protection is disabled, your Amazon EC2 instances are not protected from termination when the Auto Scaling group scales in.

autoScalingGroupArn :: Text

The Amazon Resource Name (ARN) that identifies the Auto Scaling group.

newAutoScalingGroupProvider

Arguments

:: Text
$sel:autoScalingGroupArn:AutoScalingGroupProvider'

-> AutoScalingGroupProvider

Create a value of AutoScalingGroupProvider with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:managedScaling:AutoScalingGroupProvider', autoScalingGroupProvider_managedScaling - The managed scaling settings for the Auto Scaling group capacity provider.

$sel:managedTerminationProtection:AutoScalingGroupProvider', autoScalingGroupProvider_managedTerminationProtection - The managed termination protection setting to use for the Auto Scaling group capacity provider. This determines whether the Auto Scaling group has managed termination protection.

When using managed termination protection, managed scaling must also be used otherwise managed termination protection will not work.

When managed termination protection is enabled, Amazon ECS prevents the Amazon EC2 instances in an Auto Scaling group that contain tasks from being terminated during a scale-in action. The Auto Scaling group and each instance in the Auto Scaling group must have instance protection from scale-in actions enabled as well. For more information, see Instance Protection in the Auto Scaling User Guide.

When managed termination protection is disabled, your Amazon EC2 instances are not protected from termination when the Auto Scaling group scales in.

$sel:autoScalingGroupArn:AutoScalingGroupProvider', autoScalingGroupProvider_autoScalingGroupArn - The Amazon Resource Name (ARN) that identifies the Auto Scaling group.

autoScalingGroupProvider_managedScaling :: Lens' AutoScalingGroupProvider (Maybe ManagedScaling)

The managed scaling settings for the Auto Scaling group capacity provider.

autoScalingGroupProvider_managedTerminationProtection :: Lens' AutoScalingGroupProvider (Maybe ManagedTerminationProtection)

The managed termination protection setting to use for the Auto Scaling group capacity provider. This determines whether the Auto Scaling group has managed termination protection.

When using managed termination protection, managed scaling must also be used otherwise managed termination protection will not work.

When managed termination protection is enabled, Amazon ECS prevents the Amazon EC2 instances in an Auto Scaling group that contain tasks from being terminated during a scale-in action. The Auto Scaling group and each instance in the Auto Scaling group must have instance protection from scale-in actions enabled as well. For more information, see Instance Protection in the Auto Scaling User Guide.

When managed termination protection is disabled, your Amazon EC2 instances are not protected from termination when the Auto Scaling group scales in.

autoScalingGroupProvider_autoScalingGroupArn :: Lens' AutoScalingGroupProvider Text

The Amazon Resource Name (ARN) that identifies the Auto Scaling group.