Amazonka.ECS.Types.CapacityProvider

Description

Documentation

data CapacityProvider

The details of a capacity provider.

See: newCapacityProvider smart constructor.

Constructors

CapacityProvider'

Fields

status :: Maybe CapacityProviderStatus

The current status of the capacity provider. Only capacity providers in an ACTIVE state can be used in a cluster. When a capacity provider is successfully deleted, it will have an INACTIVE status.

updateStatusReason :: Maybe Text

The update status reason. This provides further details about the update status for the capacity provider.

autoScalingGroupProvider :: Maybe AutoScalingGroupProvider

The Auto Scaling group settings for the capacity provider.

name :: Maybe Text

The name of the capacity provider.

updateStatus :: Maybe CapacityProviderUpdateStatus

The update status of the capacity provider. The following are the possible states that will be returned.

DELETE_IN_PROGRESS

The capacity provider is in the process of being deleted.

DELETE_COMPLETE

The capacity provider has been successfully deleted and will have an INACTIVE status.

DELETE_FAILED

The capacity provider was unable to be deleted. The update status reason will provide further details about why the delete failed.

capacityProviderArn :: Maybe Text

The Amazon Resource Name (ARN) that identifies the capacity provider.

tags :: Maybe [Tag]

The metadata that you apply to the capacity provider to help you categorize and organize it. Each tag consists of a key and an optional value, both of which you define.

The following basic restrictions apply to tags:

Maximum number of tags per resource - 50

For each resource, each tag key must be unique, and each tag key can have only one value.

Maximum key length - 128 Unicode characters in UTF-8

Maximum value length - 256 Unicode characters in UTF-8

If your tagging schema is used across multiple services and resources, remember that other services may have restrictions on allowed characters. Generally allowed characters are: letters, numbers, and spaces representable in UTF-8, and the following characters: + - = . _ : / @.

Tag keys and values are case-sensitive.

Do not use aws:, AWS:, or any upper or lowercase combination of such as a prefix for either keys or values as it is reserved for Amazon Web Services use. You cannot edit or delete tag keys or values with this prefix. Tags with this prefix do not count against your tags per resource limit.

newCapacityProvider :: CapacityProvider

Create a value of CapacityProvider with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:status:CapacityProvider', capacityProvider_status - The current status of the capacity provider. Only capacity providers in an ACTIVE state can be used in a cluster. When a capacity provider is successfully deleted, it will have an INACTIVE status.

$sel:updateStatusReason:CapacityProvider', capacityProvider_updateStatusReason - The update status reason. This provides further details about the update status for the capacity provider.

$sel:autoScalingGroupProvider:CapacityProvider', capacityProvider_autoScalingGroupProvider - The Auto Scaling group settings for the capacity provider.

$sel:name:CapacityProvider', capacityProvider_name - The name of the capacity provider.

$sel:updateStatus:CapacityProvider', capacityProvider_updateStatus - The update status of the capacity provider. The following are the possible states that will be returned.

DELETE_IN_PROGRESS

The capacity provider is in the process of being deleted.

DELETE_COMPLETE

The capacity provider has been successfully deleted and will have an INACTIVE status.

DELETE_FAILED

The capacity provider was unable to be deleted. The update status reason will provide further details about why the delete failed.

$sel:capacityProviderArn:CapacityProvider', capacityProvider_capacityProviderArn - The Amazon Resource Name (ARN) that identifies the capacity provider.

$sel:tags:CapacityProvider', capacityProvider_tags - The metadata that you apply to the capacity provider to help you categorize and organize it. Each tag consists of a key and an optional value, both of which you define.

The following basic restrictions apply to tags:

Maximum number of tags per resource - 50

For each resource, each tag key must be unique, and each tag key can have only one value.

Maximum key length - 128 Unicode characters in UTF-8

Maximum value length - 256 Unicode characters in UTF-8

If your tagging schema is used across multiple services and resources, remember that other services may have restrictions on allowed characters. Generally allowed characters are: letters, numbers, and spaces representable in UTF-8, and the following characters: + - = . _ : / @.

Tag keys and values are case-sensitive.

Do not use aws:, AWS:, or any upper or lowercase combination of such as a prefix for either keys or values as it is reserved for Amazon Web Services use. You cannot edit or delete tag keys or values with this prefix. Tags with this prefix do not count against your tags per resource limit.

capacityProvider_status :: Lens' CapacityProvider (Maybe CapacityProviderStatus)

The current status of the capacity provider. Only capacity providers in an ACTIVE state can be used in a cluster. When a capacity provider is successfully deleted, it will have an INACTIVE status.

capacityProvider_updateStatusReason :: Lens' CapacityProvider (Maybe Text)

The update status reason. This provides further details about the update status for the capacity provider.

capacityProvider_autoScalingGroupProvider :: Lens' CapacityProvider (Maybe AutoScalingGroupProvider)

The Auto Scaling group settings for the capacity provider.

capacityProvider_name :: Lens' CapacityProvider (Maybe Text)

The name of the capacity provider.

capacityProvider_updateStatus :: Lens' CapacityProvider (Maybe CapacityProviderUpdateStatus)

The update status of the capacity provider. The following are the possible states that will be returned.

DELETE_IN_PROGRESS

The capacity provider is in the process of being deleted.

DELETE_COMPLETE

The capacity provider has been successfully deleted and will have an INACTIVE status.

DELETE_FAILED

The capacity provider was unable to be deleted. The update status reason will provide further details about why the delete failed.

capacityProvider_capacityProviderArn :: Lens' CapacityProvider (Maybe Text)

The Amazon Resource Name (ARN) that identifies the capacity provider.

capacityProvider_tags :: Lens' CapacityProvider (Maybe [Tag])

The metadata that you apply to the capacity provider to help you categorize and organize it. Each tag consists of a key and an optional value, both of which you define.

The following basic restrictions apply to tags:

Maximum number of tags per resource - 50

For each resource, each tag key must be unique, and each tag key can have only one value.

Maximum key length - 128 Unicode characters in UTF-8

Maximum value length - 256 Unicode characters in UTF-8

If your tagging schema is used across multiple services and resources, remember that other services may have restrictions on allowed characters. Generally allowed characters are: letters, numbers, and spaces representable in UTF-8, and the following characters: + - = . _ : / @.

Tag keys and values are case-sensitive.

Do not use aws:, AWS:, or any upper or lowercase combination of such as a prefix for either keys or values as it is reserved for Amazon Web Services use. You cannot edit or delete tag keys or values with this prefix. Tags with this prefix do not count against your tags per resource limit.