Amazonka.ECS.Types.ServiceRegistry

Description

Documentation

data ServiceRegistry

Details of the service registry.

See: newServiceRegistry smart constructor.

Constructors

ServiceRegistry'

Fields

registryArn :: Maybe Text

The Amazon Resource Name (ARN) of the service registry. The currently supported service registry is Cloud Map. For more information, see CreateService.

containerName :: Maybe Text

The container name value, already specified in the task definition, to be used for your service discovery service. If the task definition that your service task specifies uses the bridge or host network mode, you must specify a containerName and containerPort combination from the task definition. If the task definition that your service task specifies uses the awsvpc network mode and a type SRV DNS record is used, you must specify either a containerName and containerPort combination or a port value, but not both.

containerPort :: Maybe Int

The port value, already specified in the task definition, to be used for your service discovery service. If the task definition your service task specifies uses the bridge or host network mode, you must specify a containerName and containerPort combination from the task definition. If the task definition your service task specifies uses the awsvpc network mode and a type SRV DNS record is used, you must specify either a containerName and containerPort combination or a port value, but not both.

port :: Maybe Int

The port value used if your service discovery service specified an SRV record. This field may be used if both the awsvpc network mode and SRV records are used.

newServiceRegistry :: ServiceRegistry

Create a value of ServiceRegistry with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:registryArn:ServiceRegistry', serviceRegistry_registryArn - The Amazon Resource Name (ARN) of the service registry. The currently supported service registry is Cloud Map. For more information, see CreateService.

$sel:containerName:ServiceRegistry', serviceRegistry_containerName - The container name value, already specified in the task definition, to be used for your service discovery service. If the task definition that your service task specifies uses the bridge or host network mode, you must specify a containerName and containerPort combination from the task definition. If the task definition that your service task specifies uses the awsvpc network mode and a type SRV DNS record is used, you must specify either a containerName and containerPort combination or a port value, but not both.

$sel:containerPort:ServiceRegistry', serviceRegistry_containerPort - The port value, already specified in the task definition, to be used for your service discovery service. If the task definition your service task specifies uses the bridge or host network mode, you must specify a containerName and containerPort combination from the task definition. If the task definition your service task specifies uses the awsvpc network mode and a type SRV DNS record is used, you must specify either a containerName and containerPort combination or a port value, but not both.

$sel:port:ServiceRegistry', serviceRegistry_port - The port value used if your service discovery service specified an SRV record. This field may be used if both the awsvpc network mode and SRV records are used.

serviceRegistry_registryArn :: Lens' ServiceRegistry (Maybe Text)

The Amazon Resource Name (ARN) of the service registry. The currently supported service registry is Cloud Map. For more information, see CreateService.

serviceRegistry_containerName :: Lens' ServiceRegistry (Maybe Text)

The container name value, already specified in the task definition, to be used for your service discovery service. If the task definition that your service task specifies uses the bridge or host network mode, you must specify a containerName and containerPort combination from the task definition. If the task definition that your service task specifies uses the awsvpc network mode and a type SRV DNS record is used, you must specify either a containerName and containerPort combination or a port value, but not both.

serviceRegistry_containerPort :: Lens' ServiceRegistry (Maybe Int)

The port value, already specified in the task definition, to be used for your service discovery service. If the task definition your service task specifies uses the bridge or host network mode, you must specify a containerName and containerPort combination from the task definition. If the task definition your service task specifies uses the awsvpc network mode and a type SRV DNS record is used, you must specify either a containerName and containerPort combination or a port value, but not both.

serviceRegistry_port :: Lens' ServiceRegistry (Maybe Int)

The port value used if your service discovery service specified an SRV record. This field may be used if both the awsvpc network mode and SRV records are used.