Amazonka.ECS.Types.Task

Description

Documentation

data Task

Details on a task in a cluster.

See: newTask smart constructor.

Constructors

Task'

Fields

stoppedAt :: Maybe POSIX

The Unix timestamp for when the task was stopped (the task transitioned from the RUNNING state to the STOPPED state).

desiredStatus :: Maybe Text

The desired status of the task. For more information, see Task Lifecycle.

overrides :: Maybe TaskOverride

One or more container overrides.

inferenceAccelerators :: Maybe [InferenceAccelerator]

The Elastic Inference accelerator associated with the task.

clusterArn :: Maybe Text

The ARN of the cluster that hosts the task.

group' :: Maybe Text

The name of the task group associated with the task.

attachments :: Maybe [Attachment]

The Elastic Network Adapter associated with the task if the task uses the awsvpc network mode.

createdAt :: Maybe POSIX

The Unix timestamp for when the task was created (the task entered the PENDING state).

stopCode :: Maybe TaskStopCode

The stop code indicating why a task was stopped. The stoppedReason may contain additional details.

platformVersion :: Maybe Text

The platform version on which your task is running. A platform version is only specified for tasks using the Fargate launch type. If one is not specified, the LATEST platform version is used by default. For more information, see Fargate Platform Versions in the Amazon Elastic Container Service Developer Guide.

taskArn :: Maybe Text

The Amazon Resource Name (ARN) of the task.

containerInstanceArn :: Maybe Text

The ARN of the container instances that host the task.

executionStoppedAt :: Maybe POSIX

The Unix timestamp for when the task execution stopped.

ephemeralStorage :: Maybe EphemeralStorage

The ephemeral storage settings for the task.

lastStatus :: Maybe Text

The last known status of the task. For more information, see Task Lifecycle.

memory :: Maybe Text

The amount of memory (in MiB) used by the task as expressed in a task definition. It can be expressed as an integer using MiB, for example 1024. It can also be expressed as a string using GB, for example 1GB or 1 GB. String values are converted to an integer indicating the MiB when the task definition is registered.

If you are using the EC2 launch type, this field is optional.

If you are using the Fargate launch type, this field is required and you must use one of the following values, which determines your range of supported values for the cpu parameter:

512 (0.5 GB), 1024 (1 GB), 2048 (2 GB) - Available cpu values: 256 (.25 vCPU)

1024 (1 GB), 2048 (2 GB), 3072 (3 GB), 4096 (4 GB) - Available cpu values: 512 (.5 vCPU)

2048 (2 GB), 3072 (3 GB), 4096 (4 GB), 5120 (5 GB), 6144 (6 GB), 7168 (7 GB), 8192 (8 GB) - Available cpu values: 1024 (1 vCPU)

Between 4096 (4 GB) and 16384 (16 GB) in increments of 1024 (1 GB) - Available cpu values: 2048 (2 vCPU)

Between 8192 (8 GB) and 30720 (30 GB) in increments of 1024 (1 GB) - Available cpu values: 4096 (4 vCPU)

pullStoppedAt :: Maybe POSIX

The Unix timestamp for when the container image pull completed.

containers :: Maybe [Container]

The containers associated with the task.

startedAt :: Maybe POSIX

The Unix timestamp for when the task started (the task transitioned from the PENDING state to the RUNNING state).

availabilityZone :: Maybe Text

The availability zone of the task.

attributes :: Maybe [Attribute]

The attributes of the task

version :: Maybe Integer

The version counter for the task. Every time a task experiences a change that triggers a CloudWatch event, the version counter is incremented. If you are replicating your Amazon ECS task state with CloudWatch Events, you can compare the version of a task reported by the Amazon ECS API actions with the version reported in CloudWatch Events for the task (inside the detail object) to verify that the version in your event stream is current.

capacityProviderName :: Maybe Text

The capacity provider associated with the task.

startedBy :: Maybe Text

The tag specified when a task is started. If the task is started by an Amazon ECS service, then the startedBy parameter contains the deployment ID of the service that starts it.

stoppedReason :: Maybe Text

The reason that the task was stopped.

connectivity :: Maybe Connectivity

The connectivity status of a task.

stoppingAt :: Maybe POSIX

The Unix timestamp for when the task stops (transitions from the RUNNING state to STOPPED).

launchType :: Maybe LaunchType

The infrastructure on which your task is running. For more information, see Amazon ECS launch types in the Amazon Elastic Container Service Developer Guide.

taskDefinitionArn :: Maybe Text

The ARN of the task definition that creates the task.

healthStatus :: Maybe HealthStatus

The health status for the task, which is determined by the health of the essential containers in the task. If all essential containers in the task are reporting as HEALTHY, then the task status also reports as HEALTHY. If any essential containers in the task are reporting as UNHEALTHY or UNKNOWN, then the task status also reports as UNHEALTHY or UNKNOWN, accordingly.

The Amazon ECS container agent does not monitor or report on Docker health checks that are embedded in a container image (such as those specified in a parent image or from the image's Dockerfile) and not specified in the container definition. Health check parameters that are specified in a container definition override any Docker health checks that exist in the container image.

connectivityAt :: Maybe POSIX

The Unix timestamp for when the task last went into CONNECTED status.

cpu :: Maybe Text

The number of CPU units used by the task as expressed in a task definition. It can be expressed as an integer using CPU units, for example 1024. It can also be expressed as a string using vCPUs, for example 1 vCPU or 1 vcpu. String values are converted to an integer indicating the CPU units when the task definition is registered.

If you are using the EC2 launch type, this field is optional. Supported values are between 128 CPU units (0.125 vCPUs) and 10240 CPU units (10 vCPUs).

If you are using the Fargate launch type, this field is required and you must use one of the following values, which determines your range of supported values for the memory parameter:

256 (.25 vCPU) - Available memory values: 512 (0.5 GB), 1024 (1 GB), 2048 (2 GB)

512 (.5 vCPU) - Available memory values: 1024 (1 GB), 2048 (2 GB), 3072 (3 GB), 4096 (4 GB)

1024 (1 vCPU) - Available memory values: 2048 (2 GB), 3072 (3 GB), 4096 (4 GB), 5120 (5 GB), 6144 (6 GB), 7168 (7 GB), 8192 (8 GB)

2048 (2 vCPU) - Available memory values: Between 4096 (4 GB) and 16384 (16 GB) in increments of 1024 (1 GB)

4096 (4 vCPU) - Available memory values: Between 8192 (8 GB) and 30720 (30 GB) in increments of 1024 (1 GB)

enableExecuteCommand :: Maybe Bool

Whether or not execute command functionality is enabled for this task. If true, this enables execute command functionality on all containers in the task.

pullStartedAt :: Maybe POSIX

The Unix timestamp for when the container image pull began.

tags :: Maybe [Tag]

The metadata that you apply to the task to help you categorize and organize them. Each tag consists of a key and an optional value, both of which you define.

The following basic restrictions apply to tags:

Maximum number of tags per resource - 50

For each resource, each tag key must be unique, and each tag key can have only one value.

Maximum key length - 128 Unicode characters in UTF-8

Maximum value length - 256 Unicode characters in UTF-8

If your tagging schema is used across multiple services and resources, remember that other services may have restrictions on allowed characters. Generally allowed characters are: letters, numbers, and spaces representable in UTF-8, and the following characters: + - = . _ : / @.

Tag keys and values are case-sensitive.

Do not use aws:, AWS:, or any upper or lowercase combination of such as a prefix for either keys or values as it is reserved for Amazon Web Services use. You cannot edit or delete tag keys or values with this prefix. Tags with this prefix do not count against your tags per resource limit.

newTask :: Task

Create a value of Task with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:stoppedAt:Task', task_stoppedAt - The Unix timestamp for when the task was stopped (the task transitioned from the RUNNING state to the STOPPED state).

$sel:desiredStatus:Task', task_desiredStatus - The desired status of the task. For more information, see Task Lifecycle.

$sel:overrides:Task', task_overrides - One or more container overrides.

$sel:inferenceAccelerators:Task', task_inferenceAccelerators - The Elastic Inference accelerator associated with the task.

$sel:clusterArn:Task', task_clusterArn - The ARN of the cluster that hosts the task.

$sel:group':Task', task_group - The name of the task group associated with the task.

$sel:attachments:Task', task_attachments - The Elastic Network Adapter associated with the task if the task uses the awsvpc network mode.

$sel:createdAt:Task', task_createdAt - The Unix timestamp for when the task was created (the task entered the PENDING state).

$sel:stopCode:Task', task_stopCode - The stop code indicating why a task was stopped. The stoppedReason may contain additional details.

$sel:platformVersion:Task', task_platformVersion - The platform version on which your task is running. A platform version is only specified for tasks using the Fargate launch type. If one is not specified, the LATEST platform version is used by default. For more information, see Fargate Platform Versions in the Amazon Elastic Container Service Developer Guide.

$sel:taskArn:Task', task_taskArn - The Amazon Resource Name (ARN) of the task.

$sel:containerInstanceArn:Task', task_containerInstanceArn - The ARN of the container instances that host the task.

$sel:executionStoppedAt:Task', task_executionStoppedAt - The Unix timestamp for when the task execution stopped.

$sel:ephemeralStorage:Task', task_ephemeralStorage - The ephemeral storage settings for the task.

$sel:lastStatus:Task', task_lastStatus - The last known status of the task. For more information, see Task Lifecycle.

$sel:memory:Task', task_memory - The amount of memory (in MiB) used by the task as expressed in a task definition. It can be expressed as an integer using MiB, for example 1024. It can also be expressed as a string using GB, for example 1GB or 1 GB. String values are converted to an integer indicating the MiB when the task definition is registered.

If you are using the EC2 launch type, this field is optional.

If you are using the Fargate launch type, this field is required and you must use one of the following values, which determines your range of supported values for the cpu parameter:

512 (0.5 GB), 1024 (1 GB), 2048 (2 GB) - Available cpu values: 256 (.25 vCPU)

1024 (1 GB), 2048 (2 GB), 3072 (3 GB), 4096 (4 GB) - Available cpu values: 512 (.5 vCPU)

2048 (2 GB), 3072 (3 GB), 4096 (4 GB), 5120 (5 GB), 6144 (6 GB), 7168 (7 GB), 8192 (8 GB) - Available cpu values: 1024 (1 vCPU)

Between 4096 (4 GB) and 16384 (16 GB) in increments of 1024 (1 GB) - Available cpu values: 2048 (2 vCPU)

Between 8192 (8 GB) and 30720 (30 GB) in increments of 1024 (1 GB) - Available cpu values: 4096 (4 vCPU)

$sel:pullStoppedAt:Task', task_pullStoppedAt - The Unix timestamp for when the container image pull completed.

$sel:containers:Task', task_containers - The containers associated with the task.

$sel:startedAt:Task', task_startedAt - The Unix timestamp for when the task started (the task transitioned from the PENDING state to the RUNNING state).

$sel:availabilityZone:Task', task_availabilityZone - The availability zone of the task.

$sel:attributes:Task', task_attributes - The attributes of the task

$sel:version:Task', task_version - The version counter for the task. Every time a task experiences a change that triggers a CloudWatch event, the version counter is incremented. If you are replicating your Amazon ECS task state with CloudWatch Events, you can compare the version of a task reported by the Amazon ECS API actions with the version reported in CloudWatch Events for the task (inside the detail object) to verify that the version in your event stream is current.

$sel:capacityProviderName:Task', task_capacityProviderName - The capacity provider associated with the task.

$sel:startedBy:Task', task_startedBy - The tag specified when a task is started. If the task is started by an Amazon ECS service, then the startedBy parameter contains the deployment ID of the service that starts it.

$sel:stoppedReason:Task', task_stoppedReason - The reason that the task was stopped.

$sel:connectivity:Task', task_connectivity - The connectivity status of a task.

$sel:stoppingAt:Task', task_stoppingAt - The Unix timestamp for when the task stops (transitions from the RUNNING state to STOPPED).

$sel:launchType:Task', task_launchType - The infrastructure on which your task is running. For more information, see Amazon ECS launch types in the Amazon Elastic Container Service Developer Guide.

$sel:taskDefinitionArn:Task', task_taskDefinitionArn - The ARN of the task definition that creates the task.

$sel:healthStatus:Task', task_healthStatus - The health status for the task, which is determined by the health of the essential containers in the task. If all essential containers in the task are reporting as HEALTHY, then the task status also reports as HEALTHY. If any essential containers in the task are reporting as UNHEALTHY or UNKNOWN, then the task status also reports as UNHEALTHY or UNKNOWN, accordingly.

The Amazon ECS container agent does not monitor or report on Docker health checks that are embedded in a container image (such as those specified in a parent image or from the image's Dockerfile) and not specified in the container definition. Health check parameters that are specified in a container definition override any Docker health checks that exist in the container image.

$sel:connectivityAt:Task', task_connectivityAt - The Unix timestamp for when the task last went into CONNECTED status.

$sel:cpu:Task', task_cpu - The number of CPU units used by the task as expressed in a task definition. It can be expressed as an integer using CPU units, for example 1024. It can also be expressed as a string using vCPUs, for example 1 vCPU or 1 vcpu. String values are converted to an integer indicating the CPU units when the task definition is registered.

If you are using the EC2 launch type, this field is optional. Supported values are between 128 CPU units (0.125 vCPUs) and 10240 CPU units (10 vCPUs).

If you are using the Fargate launch type, this field is required and you must use one of the following values, which determines your range of supported values for the memory parameter:

256 (.25 vCPU) - Available memory values: 512 (0.5 GB), 1024 (1 GB), 2048 (2 GB)

512 (.5 vCPU) - Available memory values: 1024 (1 GB), 2048 (2 GB), 3072 (3 GB), 4096 (4 GB)

1024 (1 vCPU) - Available memory values: 2048 (2 GB), 3072 (3 GB), 4096 (4 GB), 5120 (5 GB), 6144 (6 GB), 7168 (7 GB), 8192 (8 GB)

2048 (2 vCPU) - Available memory values: Between 4096 (4 GB) and 16384 (16 GB) in increments of 1024 (1 GB)

4096 (4 vCPU) - Available memory values: Between 8192 (8 GB) and 30720 (30 GB) in increments of 1024 (1 GB)

$sel:enableExecuteCommand:Task', task_enableExecuteCommand - Whether or not execute command functionality is enabled for this task. If true, this enables execute command functionality on all containers in the task.

$sel:pullStartedAt:Task', task_pullStartedAt - The Unix timestamp for when the container image pull began.

$sel:tags:Task', task_tags - The metadata that you apply to the task to help you categorize and organize them. Each tag consists of a key and an optional value, both of which you define.

The following basic restrictions apply to tags:

Maximum number of tags per resource - 50

For each resource, each tag key must be unique, and each tag key can have only one value.

Maximum key length - 128 Unicode characters in UTF-8

Maximum value length - 256 Unicode characters in UTF-8

If your tagging schema is used across multiple services and resources, remember that other services may have restrictions on allowed characters. Generally allowed characters are: letters, numbers, and spaces representable in UTF-8, and the following characters: + - = . _ : / @.

Tag keys and values are case-sensitive.

Do not use aws:, AWS:, or any upper or lowercase combination of such as a prefix for either keys or values as it is reserved for Amazon Web Services use. You cannot edit or delete tag keys or values with this prefix. Tags with this prefix do not count against your tags per resource limit.

task_stoppedAt :: Lens' Task (Maybe UTCTime)

The Unix timestamp for when the task was stopped (the task transitioned from the RUNNING state to the STOPPED state).

task_desiredStatus :: Lens' Task (Maybe Text)

The desired status of the task. For more information, see Task Lifecycle.

task_overrides :: Lens' Task (Maybe TaskOverride)

One or more container overrides.

task_inferenceAccelerators :: Lens' Task (Maybe [InferenceAccelerator])

The Elastic Inference accelerator associated with the task.

task_clusterArn :: Lens' Task (Maybe Text)

The ARN of the cluster that hosts the task.

task_group :: Lens' Task (Maybe Text)

The name of the task group associated with the task.

task_attachments :: Lens' Task (Maybe [Attachment])

The Elastic Network Adapter associated with the task if the task uses the awsvpc network mode.

task_createdAt :: Lens' Task (Maybe UTCTime)

The Unix timestamp for when the task was created (the task entered the PENDING state).

task_stopCode :: Lens' Task (Maybe TaskStopCode)

The stop code indicating why a task was stopped. The stoppedReason may contain additional details.

task_platformVersion :: Lens' Task (Maybe Text)

The platform version on which your task is running. A platform version is only specified for tasks using the Fargate launch type. If one is not specified, the LATEST platform version is used by default. For more information, see Fargate Platform Versions in the Amazon Elastic Container Service Developer Guide.

task_taskArn :: Lens' Task (Maybe Text)

The Amazon Resource Name (ARN) of the task.

task_containerInstanceArn :: Lens' Task (Maybe Text)

The ARN of the container instances that host the task.

task_executionStoppedAt :: Lens' Task (Maybe UTCTime)

The Unix timestamp for when the task execution stopped.

task_ephemeralStorage :: Lens' Task (Maybe EphemeralStorage)

The ephemeral storage settings for the task.

task_lastStatus :: Lens' Task (Maybe Text)

The last known status of the task. For more information, see Task Lifecycle.

task_memory :: Lens' Task (Maybe Text)

The amount of memory (in MiB) used by the task as expressed in a task definition. It can be expressed as an integer using MiB, for example 1024. It can also be expressed as a string using GB, for example 1GB or 1 GB. String values are converted to an integer indicating the MiB when the task definition is registered.

If you are using the EC2 launch type, this field is optional.

If you are using the Fargate launch type, this field is required and you must use one of the following values, which determines your range of supported values for the cpu parameter:

512 (0.5 GB), 1024 (1 GB), 2048 (2 GB) - Available cpu values: 256 (.25 vCPU)

1024 (1 GB), 2048 (2 GB), 3072 (3 GB), 4096 (4 GB) - Available cpu values: 512 (.5 vCPU)

2048 (2 GB), 3072 (3 GB), 4096 (4 GB), 5120 (5 GB), 6144 (6 GB), 7168 (7 GB), 8192 (8 GB) - Available cpu values: 1024 (1 vCPU)

Between 4096 (4 GB) and 16384 (16 GB) in increments of 1024 (1 GB) - Available cpu values: 2048 (2 vCPU)

Between 8192 (8 GB) and 30720 (30 GB) in increments of 1024 (1 GB) - Available cpu values: 4096 (4 vCPU)

task_pullStoppedAt :: Lens' Task (Maybe UTCTime)

The Unix timestamp for when the container image pull completed.

task_containers :: Lens' Task (Maybe [Container])

The containers associated with the task.

task_startedAt :: Lens' Task (Maybe UTCTime)

The Unix timestamp for when the task started (the task transitioned from the PENDING state to the RUNNING state).

task_availabilityZone :: Lens' Task (Maybe Text)

The availability zone of the task.

task_attributes :: Lens' Task (Maybe [Attribute])

The attributes of the task

task_version :: Lens' Task (Maybe Integer)

The version counter for the task. Every time a task experiences a change that triggers a CloudWatch event, the version counter is incremented. If you are replicating your Amazon ECS task state with CloudWatch Events, you can compare the version of a task reported by the Amazon ECS API actions with the version reported in CloudWatch Events for the task (inside the detail object) to verify that the version in your event stream is current.

task_capacityProviderName :: Lens' Task (Maybe Text)

The capacity provider associated with the task.

task_startedBy :: Lens' Task (Maybe Text)

The tag specified when a task is started. If the task is started by an Amazon ECS service, then the startedBy parameter contains the deployment ID of the service that starts it.

task_stoppedReason :: Lens' Task (Maybe Text)

The reason that the task was stopped.

task_connectivity :: Lens' Task (Maybe Connectivity)

The connectivity status of a task.

task_stoppingAt :: Lens' Task (Maybe UTCTime)

The Unix timestamp for when the task stops (transitions from the RUNNING state to STOPPED).

task_launchType :: Lens' Task (Maybe LaunchType)

The infrastructure on which your task is running. For more information, see Amazon ECS launch types in the Amazon Elastic Container Service Developer Guide.

task_taskDefinitionArn :: Lens' Task (Maybe Text)

The ARN of the task definition that creates the task.

task_healthStatus :: Lens' Task (Maybe HealthStatus)

The health status for the task, which is determined by the health of the essential containers in the task. If all essential containers in the task are reporting as HEALTHY, then the task status also reports as HEALTHY. If any essential containers in the task are reporting as UNHEALTHY or UNKNOWN, then the task status also reports as UNHEALTHY or UNKNOWN, accordingly.

The Amazon ECS container agent does not monitor or report on Docker health checks that are embedded in a container image (such as those specified in a parent image or from the image's Dockerfile) and not specified in the container definition. Health check parameters that are specified in a container definition override any Docker health checks that exist in the container image.

task_connectivityAt :: Lens' Task (Maybe UTCTime)

The Unix timestamp for when the task last went into CONNECTED status.

task_cpu :: Lens' Task (Maybe Text)

The number of CPU units used by the task as expressed in a task definition. It can be expressed as an integer using CPU units, for example 1024. It can also be expressed as a string using vCPUs, for example 1 vCPU or 1 vcpu. String values are converted to an integer indicating the CPU units when the task definition is registered.

If you are using the EC2 launch type, this field is optional. Supported values are between 128 CPU units (0.125 vCPUs) and 10240 CPU units (10 vCPUs).

If you are using the Fargate launch type, this field is required and you must use one of the following values, which determines your range of supported values for the memory parameter:

256 (.25 vCPU) - Available memory values: 512 (0.5 GB), 1024 (1 GB), 2048 (2 GB)

512 (.5 vCPU) - Available memory values: 1024 (1 GB), 2048 (2 GB), 3072 (3 GB), 4096 (4 GB)

1024 (1 vCPU) - Available memory values: 2048 (2 GB), 3072 (3 GB), 4096 (4 GB), 5120 (5 GB), 6144 (6 GB), 7168 (7 GB), 8192 (8 GB)

2048 (2 vCPU) - Available memory values: Between 4096 (4 GB) and 16384 (16 GB) in increments of 1024 (1 GB)

4096 (4 vCPU) - Available memory values: Between 8192 (8 GB) and 30720 (30 GB) in increments of 1024 (1 GB)

task_enableExecuteCommand :: Lens' Task (Maybe Bool)

Whether or not execute command functionality is enabled for this task. If true, this enables execute command functionality on all containers in the task.

task_pullStartedAt :: Lens' Task (Maybe UTCTime)

The Unix timestamp for when the container image pull began.

task_tags :: Lens' Task (Maybe [Tag])

The metadata that you apply to the task to help you categorize and organize them. Each tag consists of a key and an optional value, both of which you define.

The following basic restrictions apply to tags:

Maximum number of tags per resource - 50

For each resource, each tag key must be unique, and each tag key can have only one value.

Maximum key length - 128 Unicode characters in UTF-8

Maximum value length - 256 Unicode characters in UTF-8

If your tagging schema is used across multiple services and resources, remember that other services may have restrictions on allowed characters. Generally allowed characters are: letters, numbers, and spaces representable in UTF-8, and the following characters: + - = . _ : / @.

Tag keys and values are case-sensitive.

Do not use aws:, AWS:, or any upper or lowercase combination of such as a prefix for either keys or values as it is reserved for Amazon Web Services use. You cannot edit or delete tag keys or values with this prefix. Tags with this prefix do not count against your tags per resource limit.