Amazonka.ECS.Types.TaskDefinitionPlacementConstraint

Description

Documentation

data TaskDefinitionPlacementConstraint

An object representing a constraint on task placement in the task definition. For more information, see Task placement constraints in the Amazon Elastic Container Service Developer Guide.

Task placement constraints are not supported for tasks run on Fargate.

See: newTaskDefinitionPlacementConstraint smart constructor.

Constructors

TaskDefinitionPlacementConstraint'

Fields

expression :: Maybe Text

A cluster query language expression to apply to the constraint. For more information, see Cluster query language in the Amazon Elastic Container Service Developer Guide.

type' :: Maybe TaskDefinitionPlacementConstraintType

The type of constraint. The MemberOf constraint restricts selection to be from a group of valid candidates.

newTaskDefinitionPlacementConstraint :: TaskDefinitionPlacementConstraint

Create a value of TaskDefinitionPlacementConstraint with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:expression:TaskDefinitionPlacementConstraint', taskDefinitionPlacementConstraint_expression - A cluster query language expression to apply to the constraint. For more information, see Cluster query language in the Amazon Elastic Container Service Developer Guide.

$sel:type':TaskDefinitionPlacementConstraint', taskDefinitionPlacementConstraint_type - The type of constraint. The MemberOf constraint restricts selection to be from a group of valid candidates.

taskDefinitionPlacementConstraint_expression :: Lens' TaskDefinitionPlacementConstraint (Maybe Text)

A cluster query language expression to apply to the constraint. For more information, see Cluster query language in the Amazon Elastic Container Service Developer Guide.

taskDefinitionPlacementConstraint_type :: Lens' TaskDefinitionPlacementConstraint (Maybe TaskDefinitionPlacementConstraintType)

The type of constraint. The MemberOf constraint restricts selection to be from a group of valid candidates.