Amazonka.ECS.Types.Volume

Description

Documentation

data Volume

A data volume used in a task definition. For tasks that use the Amazon Elastic File System (Amazon EFS), specify an efsVolumeConfiguration. For Windows tasks that use Amazon FSx for Windows File Server file system, specify a fsxWindowsFileServerVolumeConfiguration. For tasks that use a Docker volume, specify a DockerVolumeConfiguration. For tasks that use a bind mount host volume, specify a host and optional sourcePath. For more information, see Using Data Volumes in Tasks.

See: newVolume smart constructor.

Constructors

Volume'

Fields

dockerVolumeConfiguration :: Maybe DockerVolumeConfiguration

This parameter is specified when you are using Docker volumes.

Windows containers only support the use of the local driver. To use bind mounts, specify the host parameter instead.

Docker volumes are not supported by tasks run on Fargate.

fsxWindowsFileServerVolumeConfiguration :: Maybe FSxWindowsFileServerVolumeConfiguration

This parameter is specified when you are using Amazon FSx for Windows File Server file system for task storage.

name :: Maybe Text

The name of the volume. Up to 255 letters (uppercase and lowercase), numbers, underscores, and hyphens are allowed. This name is referenced in the sourceVolume parameter of container definition mountPoints.

efsVolumeConfiguration :: Maybe EFSVolumeConfiguration

This parameter is specified when you are using an Amazon Elastic File System file system for task storage.

host :: Maybe HostVolumeProperties

This parameter is specified when you are using bind mount host volumes. The contents of the host parameter determine whether your bind mount host volume persists on the host container instance and where it is stored. If the host parameter is empty, then the Docker daemon assigns a host path for your data volume. However, the data is not guaranteed to persist after the containers associated with it stop running.

Windows containers can mount whole directories on the same drive as $env:ProgramData. Windows containers cannot mount directories on a different drive, and mount point cannot be across drives. For example, you can mount C:\my\path:C:\my\path and D:\:D:\, but not D:\my\path:C:\my\path or D:\:C:\my\path.

newVolume :: Volume

Create a value of Volume with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:dockerVolumeConfiguration:Volume', volume_dockerVolumeConfiguration - This parameter is specified when you are using Docker volumes.

Windows containers only support the use of the local driver. To use bind mounts, specify the host parameter instead.

Docker volumes are not supported by tasks run on Fargate.

$sel:fsxWindowsFileServerVolumeConfiguration:Volume', volume_fsxWindowsFileServerVolumeConfiguration - This parameter is specified when you are using Amazon FSx for Windows File Server file system for task storage.

$sel:name:Volume', volume_name - The name of the volume. Up to 255 letters (uppercase and lowercase), numbers, underscores, and hyphens are allowed. This name is referenced in the sourceVolume parameter of container definition mountPoints.

$sel:efsVolumeConfiguration:Volume', volume_efsVolumeConfiguration - This parameter is specified when you are using an Amazon Elastic File System file system for task storage.

$sel:host:Volume', volume_host - This parameter is specified when you are using bind mount host volumes. The contents of the host parameter determine whether your bind mount host volume persists on the host container instance and where it is stored. If the host parameter is empty, then the Docker daemon assigns a host path for your data volume. However, the data is not guaranteed to persist after the containers associated with it stop running.

Windows containers can mount whole directories on the same drive as $env:ProgramData. Windows containers cannot mount directories on a different drive, and mount point cannot be across drives. For example, you can mount C:\my\path:C:\my\path and D:\:D:\, but not D:\my\path:C:\my\path or D:\:C:\my\path.

volume_dockerVolumeConfiguration :: Lens' Volume (Maybe DockerVolumeConfiguration)

This parameter is specified when you are using Docker volumes.

Windows containers only support the use of the local driver. To use bind mounts, specify the host parameter instead.

Docker volumes are not supported by tasks run on Fargate.

volume_fsxWindowsFileServerVolumeConfiguration :: Lens' Volume (Maybe FSxWindowsFileServerVolumeConfiguration)

This parameter is specified when you are using Amazon FSx for Windows File Server file system for task storage.

volume_name :: Lens' Volume (Maybe Text)

The name of the volume. Up to 255 letters (uppercase and lowercase), numbers, underscores, and hyphens are allowed. This name is referenced in the sourceVolume parameter of container definition mountPoints.

volume_efsVolumeConfiguration :: Lens' Volume (Maybe EFSVolumeConfiguration)

This parameter is specified when you are using an Amazon Elastic File System file system for task storage.

volume_host :: Lens' Volume (Maybe HostVolumeProperties)

This parameter is specified when you are using bind mount host volumes. The contents of the host parameter determine whether your bind mount host volume persists on the host container instance and where it is stored. If the host parameter is empty, then the Docker daemon assigns a host path for your data volume. However, the data is not guaranteed to persist after the containers associated with it stop running.

Windows containers can mount whole directories on the same drive as $env:ProgramData. Windows containers cannot mount directories on a different drive, and mount point cannot be across drives. For example, you can mount C:\my\path:C:\my\path and D:\:D:\, but not D:\my\path:C:\my\path or D:\:C:\my\path.