Amazonka.EFS.CreateFileSystem

Description

Creates a new, empty file system. The operation requires a creation token in the request that Amazon EFS uses to ensure idempotent creation (calling the operation with same creation token has no effect). If a file system does not currently exist that is owned by the caller's Amazon Web Services account with the specified creation token, this operation does the following:

Creates a new, empty file system. The file system will have an Amazon EFS assigned ID, and an initial lifecycle state creating.

Returns with the description of the created file system.

Otherwise, this operation returns a FileSystemAlreadyExists error with the ID of the existing file system.

For basic use cases, you can use a randomly generated UUID for the creation token.

The idempotent operation allows you to retry a CreateFileSystem call without risk of creating an extra file system. This can happen when an initial call fails in a way that leaves it uncertain whether or not a file system was actually created. An example might be that a transport level timeout occurred or your connection was reset. As long as you use the same creation token, if the initial call had succeeded in creating a file system, the client can learn of its existence from the FileSystemAlreadyExists error.

For more information, see Creating a file system in the Amazon EFS User Guide.

The CreateFileSystem call returns while the file system's lifecycle state is still creating. You can check the file system creation status by calling the DescribeFileSystems operation, which among other things returns the file system state.

This operation accepts an optional PerformanceMode parameter that you choose for your file system. We recommend generalPurpose performance mode for most file systems. File systems using the maxIO performance mode can scale to higher levels of aggregate throughput and operations per second with a tradeoff of slightly higher latencies for most file operations. The performance mode can't be changed after the file system has been created. For more information, see Amazon EFS performance modes.

You can set the throughput mode for the file system using the ThroughputMode parameter.

After the file system is fully created, Amazon EFS sets its lifecycle state to available, at which point you can create one or more mount targets for the file system in your VPC. For more information, see CreateMountTarget. You mount your Amazon EFS file system on an EC2 instances in your VPC by using the mount target. For more information, see Amazon EFS: How it Works.

This operation requires permissions for the elasticfilesystem:CreateFileSystem action.

Creating a Request

data CreateFileSystem

See: newCreateFileSystem smart constructor.

Constructors

CreateFileSystem'

Fields

provisionedThroughputInMibps :: Maybe Double

The throughput, measured in MiB/s, that you want to provision for a file system that you're creating. Valid values are 1-1024. Required if ThroughputMode is set to provisioned. The upper limit for throughput is 1024 MiB/s. To increase this limit, contact Amazon Web Services Support. For more information, see Amazon EFS quotas that you can increase in the Amazon EFS User Guide.

availabilityZoneName :: Maybe Text

Used to create a file system that uses One Zone storage classes. It specifies the Amazon Web Services Availability Zone in which to create the file system. Use the format us-east-1a to specify the Availability Zone. For more information about One Zone storage classes, see Using EFS storage classes in the Amazon EFS User Guide.

One Zone storage classes are not available in all Availability Zones in Amazon Web Services Regions where Amazon EFS is available.

performanceMode :: Maybe PerformanceMode

The performance mode of the file system. We recommend generalPurpose performance mode for most file systems. File systems using the maxIO performance mode can scale to higher levels of aggregate throughput and operations per second with a tradeoff of slightly higher latencies for most file operations. The performance mode can't be changed after the file system has been created.

The maxIO mode is not supported on file systems using One Zone storage classes.

backup :: Maybe Bool

Specifies whether automatic backups are enabled on the file system that you are creating. Set the value to true to enable automatic backups. If you are creating a file system that uses One Zone storage classes, automatic backups are enabled by default. For more information, see Automatic backups in the Amazon EFS User Guide.

Default is false. However, if you specify an AvailabilityZoneName, the default is true.

Backup is not available in all Amazon Web Services Regionswhere Amazon EFS is available.

encrypted :: Maybe Bool

A Boolean value that, if true, creates an encrypted file system. When creating an encrypted file system, you have the option of specifying CreateFileSystemRequest$KmsKeyId for an existing Key Management Service (KMS customer master key (CMK). If you don't specify a CMK, then the default CMK for Amazon EFS, /aws/elasticfilesystem, is used to protect the encrypted file system.

throughputMode :: Maybe ThroughputMode

Specifies the throughput mode for the file system, either bursting or provisioned. If you set ThroughputMode to provisioned, you must also set a value for ProvisionedThroughputInMibps. After you create the file system, you can decrease your file system's throughput in Provisioned Throughput mode or change between the throughput modes, as long as it’s been more than 24 hours since the last decrease or throughput mode change. For more information, see Specifying throughput with provisioned mode in the Amazon EFS User Guide.

Default is bursting.

kmsKeyId :: Maybe Text

The ID of the KMS CMK that you want to use to protect the encrypted file system. This parameter is only required if you want to use a non-default KMS key. If this parameter is not specified, the default CMK for Amazon EFS is used. This ID can be in one of the following formats:

Key ID - A unique identifier of the key, for example 1234abcd-12ab-34cd-56ef-1234567890ab.

ARN - An Amazon Resource Name (ARN) for the key, for example arn:aws:kms:us-west-2:111122223333:key/1234abcd-12ab-34cd-56ef-1234567890ab.

Key alias - A previously created display name for a key, for example alias/projectKey1.

Key alias ARN - An ARN for a key alias, for example arn:aws:kms:us-west-2:444455556666:alias/projectKey1.

If KmsKeyId is specified, the CreateFileSystemRequest$Encrypted parameter must be set to true.

EFS accepts only symmetric KMS keys. You cannot use asymmetric KMS keys with EFS file systems.

tags :: Maybe [Tag]

Use to create one or more tags associated with the file system. Each tag is a user-defined key-value pair. Name your file system on creation by including a "Key":"Name","Value":"{value}" key-value pair. Each key must be unique. For more information, see Tagging Amazon Web Services resources in the Amazon Web Services General Reference Guide.

creationToken :: Text

A string of up to 64 ASCII characters. Amazon EFS uses this to ensure idempotent creation.

newCreateFileSystem

Arguments

:: Text
$sel:creationToken:CreateFileSystem'

-> CreateFileSystem

Create a value of CreateFileSystem with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:provisionedThroughputInMibps:CreateFileSystem', createFileSystem_provisionedThroughputInMibps - The throughput, measured in MiB/s, that you want to provision for a file system that you're creating. Valid values are 1-1024. Required if ThroughputMode is set to provisioned. The upper limit for throughput is 1024 MiB/s. To increase this limit, contact Amazon Web Services Support. For more information, see Amazon EFS quotas that you can increase in the Amazon EFS User Guide.

$sel:availabilityZoneName:CreateFileSystem', createFileSystem_availabilityZoneName - Used to create a file system that uses One Zone storage classes. It specifies the Amazon Web Services Availability Zone in which to create the file system. Use the format us-east-1a to specify the Availability Zone. For more information about One Zone storage classes, see Using EFS storage classes in the Amazon EFS User Guide.

One Zone storage classes are not available in all Availability Zones in Amazon Web Services Regions where Amazon EFS is available.

$sel:performanceMode:CreateFileSystem', createFileSystem_performanceMode - The performance mode of the file system. We recommend generalPurpose performance mode for most file systems. File systems using the maxIO performance mode can scale to higher levels of aggregate throughput and operations per second with a tradeoff of slightly higher latencies for most file operations. The performance mode can't be changed after the file system has been created.

The maxIO mode is not supported on file systems using One Zone storage classes.

$sel:backup:CreateFileSystem', createFileSystem_backup - Specifies whether automatic backups are enabled on the file system that you are creating. Set the value to true to enable automatic backups. If you are creating a file system that uses One Zone storage classes, automatic backups are enabled by default. For more information, see Automatic backups in the Amazon EFS User Guide.

Default is false. However, if you specify an AvailabilityZoneName, the default is true.

Backup is not available in all Amazon Web Services Regionswhere Amazon EFS is available.

$sel:encrypted:CreateFileSystem', createFileSystem_encrypted - A Boolean value that, if true, creates an encrypted file system. When creating an encrypted file system, you have the option of specifying CreateFileSystemRequest$KmsKeyId for an existing Key Management Service (KMS customer master key (CMK). If you don't specify a CMK, then the default CMK for Amazon EFS, /aws/elasticfilesystem, is used to protect the encrypted file system.

$sel:throughputMode:CreateFileSystem', createFileSystem_throughputMode - Specifies the throughput mode for the file system, either bursting or provisioned. If you set ThroughputMode to provisioned, you must also set a value for ProvisionedThroughputInMibps. After you create the file system, you can decrease your file system's throughput in Provisioned Throughput mode or change between the throughput modes, as long as it’s been more than 24 hours since the last decrease or throughput mode change. For more information, see Specifying throughput with provisioned mode in the Amazon EFS User Guide.

Default is bursting.

$sel:kmsKeyId:CreateFileSystem', createFileSystem_kmsKeyId - The ID of the KMS CMK that you want to use to protect the encrypted file system. This parameter is only required if you want to use a non-default KMS key. If this parameter is not specified, the default CMK for Amazon EFS is used. This ID can be in one of the following formats:

Key ID - A unique identifier of the key, for example 1234abcd-12ab-34cd-56ef-1234567890ab.

ARN - An Amazon Resource Name (ARN) for the key, for example arn:aws:kms:us-west-2:111122223333:key/1234abcd-12ab-34cd-56ef-1234567890ab.

Key alias - A previously created display name for a key, for example alias/projectKey1.

Key alias ARN - An ARN for a key alias, for example arn:aws:kms:us-west-2:444455556666:alias/projectKey1.

If KmsKeyId is specified, the CreateFileSystemRequest$Encrypted parameter must be set to true.

EFS accepts only symmetric KMS keys. You cannot use asymmetric KMS keys with EFS file systems.

$sel:tags:CreateFileSystem', createFileSystem_tags - Use to create one or more tags associated with the file system. Each tag is a user-defined key-value pair. Name your file system on creation by including a "Key":"Name","Value":"{value}" key-value pair. Each key must be unique. For more information, see Tagging Amazon Web Services resources in the Amazon Web Services General Reference Guide.

$sel:creationToken:CreateFileSystem', createFileSystem_creationToken - A string of up to 64 ASCII characters. Amazon EFS uses this to ensure idempotent creation.

Request Lenses

createFileSystem_provisionedThroughputInMibps :: Lens' CreateFileSystem (Maybe Double)

The throughput, measured in MiB/s, that you want to provision for a file system that you're creating. Valid values are 1-1024. Required if ThroughputMode is set to provisioned. The upper limit for throughput is 1024 MiB/s. To increase this limit, contact Amazon Web Services Support. For more information, see Amazon EFS quotas that you can increase in the Amazon EFS User Guide.

createFileSystem_availabilityZoneName :: Lens' CreateFileSystem (Maybe Text)

Used to create a file system that uses One Zone storage classes. It specifies the Amazon Web Services Availability Zone in which to create the file system. Use the format us-east-1a to specify the Availability Zone. For more information about One Zone storage classes, see Using EFS storage classes in the Amazon EFS User Guide.

One Zone storage classes are not available in all Availability Zones in Amazon Web Services Regions where Amazon EFS is available.

createFileSystem_performanceMode :: Lens' CreateFileSystem (Maybe PerformanceMode)

The performance mode of the file system. We recommend generalPurpose performance mode for most file systems. File systems using the maxIO performance mode can scale to higher levels of aggregate throughput and operations per second with a tradeoff of slightly higher latencies for most file operations. The performance mode can't be changed after the file system has been created.

The maxIO mode is not supported on file systems using One Zone storage classes.

createFileSystem_backup :: Lens' CreateFileSystem (Maybe Bool)

Specifies whether automatic backups are enabled on the file system that you are creating. Set the value to true to enable automatic backups. If you are creating a file system that uses One Zone storage classes, automatic backups are enabled by default. For more information, see Automatic backups in the Amazon EFS User Guide.

Default is false. However, if you specify an AvailabilityZoneName, the default is true.

Backup is not available in all Amazon Web Services Regionswhere Amazon EFS is available.

createFileSystem_encrypted :: Lens' CreateFileSystem (Maybe Bool)

A Boolean value that, if true, creates an encrypted file system. When creating an encrypted file system, you have the option of specifying CreateFileSystemRequest$KmsKeyId for an existing Key Management Service (KMS customer master key (CMK). If you don't specify a CMK, then the default CMK for Amazon EFS, /aws/elasticfilesystem, is used to protect the encrypted file system.

createFileSystem_throughputMode :: Lens' CreateFileSystem (Maybe ThroughputMode)

Specifies the throughput mode for the file system, either bursting or provisioned. If you set ThroughputMode to provisioned, you must also set a value for ProvisionedThroughputInMibps. After you create the file system, you can decrease your file system's throughput in Provisioned Throughput mode or change between the throughput modes, as long as it’s been more than 24 hours since the last decrease or throughput mode change. For more information, see Specifying throughput with provisioned mode in the Amazon EFS User Guide.

Default is bursting.

createFileSystem_kmsKeyId :: Lens' CreateFileSystem (Maybe Text)

The ID of the KMS CMK that you want to use to protect the encrypted file system. This parameter is only required if you want to use a non-default KMS key. If this parameter is not specified, the default CMK for Amazon EFS is used. This ID can be in one of the following formats:

Key ID - A unique identifier of the key, for example 1234abcd-12ab-34cd-56ef-1234567890ab.

ARN - An Amazon Resource Name (ARN) for the key, for example arn:aws:kms:us-west-2:111122223333:key/1234abcd-12ab-34cd-56ef-1234567890ab.

Key alias - A previously created display name for a key, for example alias/projectKey1.

Key alias ARN - An ARN for a key alias, for example arn:aws:kms:us-west-2:444455556666:alias/projectKey1.

If KmsKeyId is specified, the CreateFileSystemRequest$Encrypted parameter must be set to true.

EFS accepts only symmetric KMS keys. You cannot use asymmetric KMS keys with EFS file systems.

createFileSystem_tags :: Lens' CreateFileSystem (Maybe [Tag])

Use to create one or more tags associated with the file system. Each tag is a user-defined key-value pair. Name your file system on creation by including a "Key":"Name","Value":"{value}" key-value pair. Each key must be unique. For more information, see Tagging Amazon Web Services resources in the Amazon Web Services General Reference Guide.

createFileSystem_creationToken :: Lens' CreateFileSystem Text

A string of up to 64 ASCII characters. Amazon EFS uses this to ensure idempotent creation.

Destructuring the Response

data FileSystemDescription

A description of the file system.

See: newFileSystemDescription smart constructor.

Constructors

FileSystemDescription'

Fields

availabilityZoneId :: Maybe Text

The unique and consistent identifier of the Availability Zone in which the file system's One Zone storage classes exist. For example, use1-az1 is an Availability Zone ID for the us-east-1 Amazon Web Services Region, and it has the same location in every Amazon Web Services account.

provisionedThroughputInMibps :: Maybe Double

The amount of provisioned throughput, measured in MiB/s, for the file system. Valid for file systems using ThroughputMode set to provisioned.

availabilityZoneName :: Maybe Text

Describes the Amazon Web Services Availability Zone in which the file system is located, and is valid only for file systems using One Zone storage classes. For more information, see Using EFS storage classes in the Amazon EFS User Guide.

fileSystemArn :: Maybe Text

The Amazon Resource Name (ARN) for the EFS file system, in the format arn:aws:elasticfilesystem:region:account-id:file-system/file-system-id . Example with sample data: arn:aws:elasticfilesystem:us-west-2:1111333322228888:file-system/fs-01234567

encrypted :: Maybe Bool

A Boolean value that, if true, indicates that the file system is encrypted.

throughputMode :: Maybe ThroughputMode

Displays the file system's throughput mode. For more information, see Throughput modes in the Amazon EFS User Guide.

kmsKeyId :: Maybe Text

The ID of an Key Management Service customer master key (CMK) that was used to protect the encrypted file system.

name :: Maybe Text

You can add tags to a file system, including a Name tag. For more information, see CreateFileSystem. If the file system has a Name tag, Amazon EFS returns the value in this field.

ownerId :: Text

The Amazon Web Services account that created the file system. If the file system was created by an IAM user, the parent account to which the user belongs is the owner.

creationToken :: Text

The opaque string specified in the request.

fileSystemId :: Text

The ID of the file system, assigned by Amazon EFS.

creationTime :: POSIX

The time that the file system was created, in seconds (since 1970-01-01T00:00:00Z).

lifeCycleState :: LifeCycleState

The lifecycle phase of the file system.

numberOfMountTargets :: Natural

The current number of mount targets that the file system has. For more information, see CreateMountTarget.

sizeInBytes :: FileSystemSize

The latest known metered size (in bytes) of data stored in the file system, in its Value field, and the time at which that size was determined in its Timestamp field. The Timestamp value is the integer number of seconds since 1970-01-01T00:00:00Z. The SizeInBytes value doesn't represent the size of a consistent snapshot of the file system, but it is eventually consistent when there are no writes to the file system. That is, SizeInBytes represents actual size only if the file system is not modified for a period longer than a couple of hours. Otherwise, the value is not the exact size that the file system was at any point in time.

performanceMode :: PerformanceMode

The performance mode of the file system.

tags :: [Tag]

The tags associated with the file system, presented as an array of Tag objects.

newFileSystemDescription

Arguments

:: Text
$sel:ownerId:FileSystemDescription'

-> Text
$sel:creationToken:FileSystemDescription'

-> Text
$sel:fileSystemId:FileSystemDescription'

-> UTCTime
$sel:creationTime:FileSystemDescription'

-> LifeCycleState
$sel:lifeCycleState:FileSystemDescription'

-> Natural
$sel:numberOfMountTargets:FileSystemDescription'

-> FileSystemSize
$sel:sizeInBytes:FileSystemDescription'

-> PerformanceMode
$sel:performanceMode:FileSystemDescription'

-> FileSystemDescription

Create a value of FileSystemDescription with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:availabilityZoneId:FileSystemDescription', fileSystemDescription_availabilityZoneId - The unique and consistent identifier of the Availability Zone in which the file system's One Zone storage classes exist. For example, use1-az1 is an Availability Zone ID for the us-east-1 Amazon Web Services Region, and it has the same location in every Amazon Web Services account.

$sel:provisionedThroughputInMibps:FileSystemDescription', fileSystemDescription_provisionedThroughputInMibps - The amount of provisioned throughput, measured in MiB/s, for the file system. Valid for file systems using ThroughputMode set to provisioned.

$sel:availabilityZoneName:FileSystemDescription', fileSystemDescription_availabilityZoneName - Describes the Amazon Web Services Availability Zone in which the file system is located, and is valid only for file systems using One Zone storage classes. For more information, see Using EFS storage classes in the Amazon EFS User Guide.

$sel:fileSystemArn:FileSystemDescription', fileSystemDescription_fileSystemArn - The Amazon Resource Name (ARN) for the EFS file system, in the format arn:aws:elasticfilesystem:region:account-id:file-system/file-system-id . Example with sample data: arn:aws:elasticfilesystem:us-west-2:1111333322228888:file-system/fs-01234567

$sel:encrypted:FileSystemDescription', fileSystemDescription_encrypted - A Boolean value that, if true, indicates that the file system is encrypted.

$sel:throughputMode:FileSystemDescription', fileSystemDescription_throughputMode - Displays the file system's throughput mode. For more information, see Throughput modes in the Amazon EFS User Guide.

$sel:kmsKeyId:FileSystemDescription', fileSystemDescription_kmsKeyId - The ID of an Key Management Service customer master key (CMK) that was used to protect the encrypted file system.

$sel:name:FileSystemDescription', fileSystemDescription_name - You can add tags to a file system, including a Name tag. For more information, see CreateFileSystem. If the file system has a Name tag, Amazon EFS returns the value in this field.

$sel:ownerId:FileSystemDescription', fileSystemDescription_ownerId - The Amazon Web Services account that created the file system. If the file system was created by an IAM user, the parent account to which the user belongs is the owner.

$sel:creationToken:FileSystemDescription', fileSystemDescription_creationToken - The opaque string specified in the request.

$sel:fileSystemId:FileSystemDescription', fileSystemDescription_fileSystemId - The ID of the file system, assigned by Amazon EFS.

$sel:creationTime:FileSystemDescription', fileSystemDescription_creationTime - The time that the file system was created, in seconds (since 1970-01-01T00:00:00Z).

$sel:lifeCycleState:FileSystemDescription', fileSystemDescription_lifeCycleState - The lifecycle phase of the file system.

$sel:numberOfMountTargets:FileSystemDescription', fileSystemDescription_numberOfMountTargets - The current number of mount targets that the file system has. For more information, see CreateMountTarget.

$sel:sizeInBytes:FileSystemDescription', fileSystemDescription_sizeInBytes - The latest known metered size (in bytes) of data stored in the file system, in its Value field, and the time at which that size was determined in its Timestamp field. The Timestamp value is the integer number of seconds since 1970-01-01T00:00:00Z. The SizeInBytes value doesn't represent the size of a consistent snapshot of the file system, but it is eventually consistent when there are no writes to the file system. That is, SizeInBytes represents actual size only if the file system is not modified for a period longer than a couple of hours. Otherwise, the value is not the exact size that the file system was at any point in time.

$sel:performanceMode:FileSystemDescription', fileSystemDescription_performanceMode - The performance mode of the file system.

$sel:tags:FileSystemDescription', fileSystemDescription_tags - The tags associated with the file system, presented as an array of Tag objects.

Response Lenses

fileSystemDescription_availabilityZoneId :: Lens' FileSystemDescription (Maybe Text)

The unique and consistent identifier of the Availability Zone in which the file system's One Zone storage classes exist. For example, use1-az1 is an Availability Zone ID for the us-east-1 Amazon Web Services Region, and it has the same location in every Amazon Web Services account.

fileSystemDescription_provisionedThroughputInMibps :: Lens' FileSystemDescription (Maybe Double)

The amount of provisioned throughput, measured in MiB/s, for the file system. Valid for file systems using ThroughputMode set to provisioned.

fileSystemDescription_availabilityZoneName :: Lens' FileSystemDescription (Maybe Text)

Describes the Amazon Web Services Availability Zone in which the file system is located, and is valid only for file systems using One Zone storage classes. For more information, see Using EFS storage classes in the Amazon EFS User Guide.

fileSystemDescription_fileSystemArn :: Lens' FileSystemDescription (Maybe Text)

The Amazon Resource Name (ARN) for the EFS file system, in the format arn:aws:elasticfilesystem:region:account-id:file-system/file-system-id . Example with sample data: arn:aws:elasticfilesystem:us-west-2:1111333322228888:file-system/fs-01234567

fileSystemDescription_encrypted :: Lens' FileSystemDescription (Maybe Bool)

A Boolean value that, if true, indicates that the file system is encrypted.

fileSystemDescription_throughputMode :: Lens' FileSystemDescription (Maybe ThroughputMode)

Displays the file system's throughput mode. For more information, see Throughput modes in the Amazon EFS User Guide.

fileSystemDescription_kmsKeyId :: Lens' FileSystemDescription (Maybe Text)

The ID of an Key Management Service customer master key (CMK) that was used to protect the encrypted file system.

fileSystemDescription_name :: Lens' FileSystemDescription (Maybe Text)

You can add tags to a file system, including a Name tag. For more information, see CreateFileSystem. If the file system has a Name tag, Amazon EFS returns the value in this field.

fileSystemDescription_ownerId :: Lens' FileSystemDescription Text

The Amazon Web Services account that created the file system. If the file system was created by an IAM user, the parent account to which the user belongs is the owner.

fileSystemDescription_creationToken :: Lens' FileSystemDescription Text

The opaque string specified in the request.

fileSystemDescription_fileSystemId :: Lens' FileSystemDescription Text

The ID of the file system, assigned by Amazon EFS.

fileSystemDescription_creationTime :: Lens' FileSystemDescription UTCTime

The time that the file system was created, in seconds (since 1970-01-01T00:00:00Z).

fileSystemDescription_lifeCycleState :: Lens' FileSystemDescription LifeCycleState

The lifecycle phase of the file system.

fileSystemDescription_numberOfMountTargets :: Lens' FileSystemDescription Natural

The current number of mount targets that the file system has. For more information, see CreateMountTarget.

fileSystemDescription_sizeInBytes :: Lens' FileSystemDescription FileSystemSize

The latest known metered size (in bytes) of data stored in the file system, in its Value field, and the time at which that size was determined in its Timestamp field. The Timestamp value is the integer number of seconds since 1970-01-01T00:00:00Z. The SizeInBytes value doesn't represent the size of a consistent snapshot of the file system, but it is eventually consistent when there are no writes to the file system. That is, SizeInBytes represents actual size only if the file system is not modified for a period longer than a couple of hours. Otherwise, the value is not the exact size that the file system was at any point in time.

fileSystemDescription_performanceMode :: Lens' FileSystemDescription PerformanceMode

The performance mode of the file system.

fileSystemDescription_tags :: Lens' FileSystemDescription [Tag]

The tags associated with the file system, presented as an array of Tag objects.