Amazonka.EFS.DeleteMountTarget

Description

Deletes the specified mount target.

This operation forcibly breaks any mounts of the file system by using the mount target that is being deleted, which might disrupt instances or applications using those mounts. To avoid applications getting cut off abruptly, you might consider unmounting any mounts of the mount target, if feasible. The operation also deletes the associated network interface. Uncommitted writes might be lost, but breaking a mount target using this operation does not corrupt the file system itself. The file system you created remains. You can mount an EC2 instance in your VPC by using another mount target.

This operation requires permissions for the following action on the file system:

elasticfilesystem:DeleteMountTarget

The DeleteMountTarget call returns while the mount target state is still deleting. You can check the mount target deletion by calling the DescribeMountTargets operation, which returns a list of mount target descriptions for the given file system.

The operation also requires permissions for the following Amazon EC2 action on the mount target's network interface:

ec2:DeleteNetworkInterface

Creating a Request

data DeleteMountTarget

See: newDeleteMountTarget smart constructor.

Constructors

DeleteMountTarget'

Fields

mountTargetId :: Text

The ID of the mount target to delete (String).

newDeleteMountTarget

Arguments

:: Text
$sel:mountTargetId:DeleteMountTarget'

-> DeleteMountTarget

Create a value of DeleteMountTarget with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:mountTargetId:DeleteMountTarget', deleteMountTarget_mountTargetId - The ID of the mount target to delete (String).

Request Lenses

deleteMountTarget_mountTargetId :: Lens' DeleteMountTarget Text

The ID of the mount target to delete (String).

Destructuring the Response

data DeleteMountTargetResponse

See: newDeleteMountTargetResponse smart constructor.

Constructors

DeleteMountTargetResponse'

newDeleteMountTargetResponse :: DeleteMountTargetResponse

Create a value of DeleteMountTargetResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.