Amazonka.EFS.DescribeMountTargets

Description

Returns the descriptions of all the current mount targets, or a specific mount target, for a file system. When requesting all of the current mount targets, the order of mount targets returned in the response is unspecified.

This operation requires permissions for the elasticfilesystem:DescribeMountTargets action, on either the file system ID that you specify in FileSystemId, or on the file system of the mount target that you specify in MountTargetId.

This operation returns paginated results.

Creating a Request

data DescribeMountTargets

See: newDescribeMountTargets smart constructor.

Constructors

DescribeMountTargets'

Fields

accessPointId :: Maybe Text

(Optional) The ID of the access point whose mount targets that you want to list. It must be included in your request if a FileSystemId or MountTargetId is not included in your request. Accepts either an access point ID or ARN as input.

fileSystemId :: Maybe Text

(Optional) ID of the file system whose mount targets you want to list (String). It must be included in your request if an AccessPointId or MountTargetId is not included. Accepts either a file system ID or ARN as input.

marker :: Maybe Text

(Optional) Opaque pagination token returned from a previous DescribeMountTargets operation (String). If present, it specifies to continue the list from where the previous returning call left off.

maxItems :: Maybe Natural

(Optional) Maximum number of mount targets to return in the response. Currently, this number is automatically set to 10, and other values are ignored. The response is paginated at 100 per page if you have more than 100 mount targets.

mountTargetId :: Maybe Text

(Optional) ID of the mount target that you want to have described (String). It must be included in your request if FileSystemId is not included. Accepts either a mount target ID or ARN as input.

newDescribeMountTargets :: DescribeMountTargets

Create a value of DescribeMountTargets with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:accessPointId:DescribeMountTargets', describeMountTargets_accessPointId - (Optional) The ID of the access point whose mount targets that you want to list. It must be included in your request if a FileSystemId or MountTargetId is not included in your request. Accepts either an access point ID or ARN as input.

$sel:fileSystemId:DescribeMountTargets', describeMountTargets_fileSystemId - (Optional) ID of the file system whose mount targets you want to list (String). It must be included in your request if an AccessPointId or MountTargetId is not included. Accepts either a file system ID or ARN as input.

$sel:marker:DescribeMountTargets', describeMountTargets_marker - (Optional) Opaque pagination token returned from a previous DescribeMountTargets operation (String). If present, it specifies to continue the list from where the previous returning call left off.

$sel:maxItems:DescribeMountTargets', describeMountTargets_maxItems - (Optional) Maximum number of mount targets to return in the response. Currently, this number is automatically set to 10, and other values are ignored. The response is paginated at 100 per page if you have more than 100 mount targets.

$sel:mountTargetId:DescribeMountTargets', describeMountTargets_mountTargetId - (Optional) ID of the mount target that you want to have described (String). It must be included in your request if FileSystemId is not included. Accepts either a mount target ID or ARN as input.

Request Lenses

describeMountTargets_accessPointId :: Lens' DescribeMountTargets (Maybe Text)

(Optional) The ID of the access point whose mount targets that you want to list. It must be included in your request if a FileSystemId or MountTargetId is not included in your request. Accepts either an access point ID or ARN as input.

describeMountTargets_fileSystemId :: Lens' DescribeMountTargets (Maybe Text)

(Optional) ID of the file system whose mount targets you want to list (String). It must be included in your request if an AccessPointId or MountTargetId is not included. Accepts either a file system ID or ARN as input.

describeMountTargets_marker :: Lens' DescribeMountTargets (Maybe Text)

(Optional) Opaque pagination token returned from a previous DescribeMountTargets operation (String). If present, it specifies to continue the list from where the previous returning call left off.

describeMountTargets_maxItems :: Lens' DescribeMountTargets (Maybe Natural)

(Optional) Maximum number of mount targets to return in the response. Currently, this number is automatically set to 10, and other values are ignored. The response is paginated at 100 per page if you have more than 100 mount targets.

describeMountTargets_mountTargetId :: Lens' DescribeMountTargets (Maybe Text)

(Optional) ID of the mount target that you want to have described (String). It must be included in your request if FileSystemId is not included. Accepts either a mount target ID or ARN as input.

Destructuring the Response

data DescribeMountTargetsResponse

See: newDescribeMountTargetsResponse smart constructor.

Constructors

DescribeMountTargetsResponse'

Fields

mountTargets :: Maybe [MountTargetDescription]

Returns the file system's mount targets as an array of MountTargetDescription objects.

marker :: Maybe Text

If the request included the Marker, the response returns that value in this field.

nextMarker :: Maybe Text

If a value is present, there are more mount targets to return. In a subsequent request, you can provide Marker in your request with this value to retrieve the next set of mount targets.

httpStatus :: Int

The response's http status code.

newDescribeMountTargetsResponse

Arguments

:: Int
$sel:httpStatus:DescribeMountTargetsResponse'

-> DescribeMountTargetsResponse

Create a value of DescribeMountTargetsResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:mountTargets:DescribeMountTargetsResponse', describeMountTargetsResponse_mountTargets - Returns the file system's mount targets as an array of MountTargetDescription objects.

$sel:marker:DescribeMountTargets', describeMountTargetsResponse_marker - If the request included the Marker, the response returns that value in this field.

$sel:nextMarker:DescribeMountTargetsResponse', describeMountTargetsResponse_nextMarker - If a value is present, there are more mount targets to return. In a subsequent request, you can provide Marker in your request with this value to retrieve the next set of mount targets.

$sel:httpStatus:DescribeMountTargetsResponse', describeMountTargetsResponse_httpStatus - The response's http status code.

Response Lenses

describeMountTargetsResponse_mountTargets :: Lens' DescribeMountTargetsResponse (Maybe [MountTargetDescription])

Returns the file system's mount targets as an array of MountTargetDescription objects.

describeMountTargetsResponse_marker :: Lens' DescribeMountTargetsResponse (Maybe Text)

If the request included the Marker, the response returns that value in this field.

describeMountTargetsResponse_nextMarker :: Lens' DescribeMountTargetsResponse (Maybe Text)

If a value is present, there are more mount targets to return. In a subsequent request, you can provide Marker in your request with this value to retrieve the next set of mount targets.

describeMountTargetsResponse_httpStatus :: Lens' DescribeMountTargetsResponse Int

The response's http status code.