Amazonka.EFS.Lens

Description

Operations

CreateAccessPoint

createAccessPoint_posixUser :: Lens' CreateAccessPoint (Maybe PosixUser)

The operating system user and group applied to all file system requests made using the access point.

createAccessPoint_rootDirectory :: Lens' CreateAccessPoint (Maybe RootDirectory)

Specifies the directory on the Amazon EFS file system that the access point exposes as the root directory of your file system to NFS clients using the access point. The clients using the access point can only access the root directory and below. If the RootDirectory > Path specified does not exist, EFS creates it and applies the CreationInfo settings when a client connects to an access point. When specifying a RootDirectory, you need to provide the Path, and the CreationInfo.

Amazon EFS creates a root directory only if you have provided the CreationInfo: OwnUid, OwnGID, and permissions for the directory. If you do not provide this information, Amazon EFS does not create the root directory. If the root directory does not exist, attempts to mount using the access point will fail.

createAccessPoint_tags :: Lens' CreateAccessPoint (Maybe [Tag])

Creates tags associated with the access point. Each tag is a key-value pair, each key must be unique. For more information, see Tagging Amazon Web Services resources in the Amazon Web Services General Reference Guide.

createAccessPoint_clientToken :: Lens' CreateAccessPoint Text

A string of up to 64 ASCII characters that Amazon EFS uses to ensure idempotent creation.

createAccessPoint_fileSystemId :: Lens' CreateAccessPoint Text

The ID of the EFS file system that the access point provides access to.

accessPointDescription_posixUser :: Lens' AccessPointDescription (Maybe PosixUser)

The full POSIX identity, including the user ID, group ID, and secondary group IDs on the access point that is used for all file operations by NFS clients using the access point.

accessPointDescription_rootDirectory :: Lens' AccessPointDescription (Maybe RootDirectory)

The directory on the Amazon EFS file system that the access point exposes as the root directory to NFS clients using the access point.

accessPointDescription_clientToken :: Lens' AccessPointDescription (Maybe Text)

The opaque string specified in the request to ensure idempotent creation.

accessPointDescription_accessPointId :: Lens' AccessPointDescription (Maybe Text)

The ID of the access point, assigned by Amazon EFS.

accessPointDescription_fileSystemId :: Lens' AccessPointDescription (Maybe Text)

The ID of the EFS file system that the access point applies to.

accessPointDescription_ownerId :: Lens' AccessPointDescription (Maybe Text)

Identified the Amazon Web Services account that owns the access point resource.

accessPointDescription_name :: Lens' AccessPointDescription (Maybe Text)

The name of the access point. This is the value of the Name tag.

accessPointDescription_accessPointArn :: Lens' AccessPointDescription (Maybe Text)

The unique Amazon Resource Name (ARN) associated with the access point.

accessPointDescription_lifeCycleState :: Lens' AccessPointDescription (Maybe LifeCycleState)

Identifies the lifecycle phase of the access point.

accessPointDescription_tags :: Lens' AccessPointDescription (Maybe [Tag])

The tags associated with the access point, presented as an array of Tag objects.

DescribeAccountPreferences

describeAccountPreferences_nextToken :: Lens' DescribeAccountPreferences (Maybe Text)

(Optional) You can use NextToken in a subsequent request to fetch the next page of Amazon Web Services account preferences if the response payload was paginated.

describeAccountPreferences_maxResults :: Lens' DescribeAccountPreferences (Maybe Natural)

(Optional) When retrieving account preferences, you can optionally specify the MaxItems parameter to limit the number of objects returned in a response. The default value is 100.

describeAccountPreferencesResponse_resourceIdPreference :: Lens' DescribeAccountPreferencesResponse (Maybe ResourceIdPreference)

Describes the resource ID preference setting for the Amazon Web Services account associated with the user making the request, in the current Amazon Web Services Region.

describeAccountPreferencesResponse_nextToken :: Lens' DescribeAccountPreferencesResponse (Maybe Text)

Present if there are more records than returned in the response. You can use the NextToken in the subsequent request to fetch the additional descriptions.

describeAccountPreferencesResponse_httpStatus :: Lens' DescribeAccountPreferencesResponse Int

The response's http status code.

DescribeMountTargets

describeMountTargets_accessPointId :: Lens' DescribeMountTargets (Maybe Text)

(Optional) The ID of the access point whose mount targets that you want to list. It must be included in your request if a FileSystemId or MountTargetId is not included in your request. Accepts either an access point ID or ARN as input.

describeMountTargets_fileSystemId :: Lens' DescribeMountTargets (Maybe Text)

(Optional) ID of the file system whose mount targets you want to list (String). It must be included in your request if an AccessPointId or MountTargetId is not included. Accepts either a file system ID or ARN as input.

describeMountTargets_marker :: Lens' DescribeMountTargets (Maybe Text)

(Optional) Opaque pagination token returned from a previous DescribeMountTargets operation (String). If present, it specifies to continue the list from where the previous returning call left off.

describeMountTargets_maxItems :: Lens' DescribeMountTargets (Maybe Natural)

(Optional) Maximum number of mount targets to return in the response. Currently, this number is automatically set to 10, and other values are ignored. The response is paginated at 100 per page if you have more than 100 mount targets.

describeMountTargets_mountTargetId :: Lens' DescribeMountTargets (Maybe Text)

(Optional) ID of the mount target that you want to have described (String). It must be included in your request if FileSystemId is not included. Accepts either a mount target ID or ARN as input.

describeMountTargetsResponse_mountTargets :: Lens' DescribeMountTargetsResponse (Maybe [MountTargetDescription])

Returns the file system's mount targets as an array of MountTargetDescription objects.

describeMountTargetsResponse_marker :: Lens' DescribeMountTargetsResponse (Maybe Text)

If the request included the Marker, the response returns that value in this field.

describeMountTargetsResponse_nextMarker :: Lens' DescribeMountTargetsResponse (Maybe Text)

If a value is present, there are more mount targets to return. In a subsequent request, you can provide Marker in your request with this value to retrieve the next set of mount targets.

describeMountTargetsResponse_httpStatus :: Lens' DescribeMountTargetsResponse Int

The response's http status code.

DeleteFileSystemPolicy

deleteFileSystemPolicy_fileSystemId :: Lens' DeleteFileSystemPolicy Text

Specifies the EFS file system for which to delete the FileSystemPolicy.

ListTagsForResource

listTagsForResource_nextToken :: Lens' ListTagsForResource (Maybe Text)

(Optional) You can use NextToken in a subsequent request to fetch the next page of access point descriptions if the response payload was paginated.

listTagsForResource_maxResults :: Lens' ListTagsForResource (Maybe Natural)

(Optional) Specifies the maximum number of tag objects to return in the response. The default value is 100.

listTagsForResource_resourceId :: Lens' ListTagsForResource Text

Specifies the EFS resource you want to retrieve tags for. You can retrieve tags for EFS file systems and access points using this API endpoint.

listTagsForResourceResponse_nextToken :: Lens' ListTagsForResourceResponse (Maybe Text)

NextToken is present if the response payload is paginated. You can use NextToken in a subsequent request to fetch the next page of access point descriptions.

listTagsForResourceResponse_tags :: Lens' ListTagsForResourceResponse (Maybe [Tag])

An array of the tags for the specified EFS resource.

listTagsForResourceResponse_httpStatus :: Lens' ListTagsForResourceResponse Int

The response's http status code.

PutFileSystemPolicy

putFileSystemPolicy_bypassPolicyLockoutSafetyCheck :: Lens' PutFileSystemPolicy (Maybe Bool)

(Optional) A flag to indicate whether to bypass the FileSystemPolicy lockout safety check. The policy lockout safety check determines whether the policy in the request will prevent the principal making the request will be locked out from making future PutFileSystemPolicy requests on the file system. Set BypassPolicyLockoutSafetyCheck to True only when you intend to prevent the principal that is making the request from making a subsequent PutFileSystemPolicy request on the file system. The default value is False.

putFileSystemPolicy_fileSystemId :: Lens' PutFileSystemPolicy Text

The ID of the EFS file system that you want to create or update the FileSystemPolicy for.

putFileSystemPolicy_policy :: Lens' PutFileSystemPolicy Text

The FileSystemPolicy that you're creating. Accepts a JSON formatted policy definition. EFS file system policies have a 20,000 character limit. To find out more about the elements that make up a file system policy, see EFS Resource-based Policies.

fileSystemPolicyDescription_fileSystemId :: Lens' FileSystemPolicyDescription (Maybe Text)

Specifies the EFS file system to which the FileSystemPolicy applies.

fileSystemPolicyDescription_policy :: Lens' FileSystemPolicyDescription (Maybe Text)

The JSON formatted FileSystemPolicy for the EFS file system.

DescribeFileSystems

describeFileSystems_fileSystemId :: Lens' DescribeFileSystems (Maybe Text)

(Optional) ID of the file system whose description you want to retrieve (String).

describeFileSystems_creationToken :: Lens' DescribeFileSystems (Maybe Text)

(Optional) Restricts the list to the file system with this creation token (String). You specify a creation token when you create an Amazon EFS file system.

describeFileSystems_marker :: Lens' DescribeFileSystems (Maybe Text)

(Optional) Opaque pagination token returned from a previous DescribeFileSystems operation (String). If present, specifies to continue the list from where the returning call had left off.

describeFileSystems_maxItems :: Lens' DescribeFileSystems (Maybe Natural)

(Optional) Specifies the maximum number of file systems to return in the response (integer). This number is automatically set to 100. The response is paginated at 100 per page if you have more than 100 file systems.

describeFileSystemsResponse_fileSystems :: Lens' DescribeFileSystemsResponse (Maybe [FileSystemDescription])

An array of file system descriptions.

describeFileSystemsResponse_marker :: Lens' DescribeFileSystemsResponse (Maybe Text)

Present if provided by caller in the request (String).

describeFileSystemsResponse_nextMarker :: Lens' DescribeFileSystemsResponse (Maybe Text)

Present if there are more file systems than returned in the response (String). You can use the NextMarker in the subsequent request to fetch the descriptions.

describeFileSystemsResponse_httpStatus :: Lens' DescribeFileSystemsResponse Int

The response's http status code.

DeleteMountTarget

deleteMountTarget_mountTargetId :: Lens' DeleteMountTarget Text

The ID of the mount target to delete (String).

PutAccountPreferences

putAccountPreferences_resourceIdType :: Lens' PutAccountPreferences ResourceIdType

Specifies the EFS resource ID preference to set for the user's Amazon Web Services account, in the current Amazon Web Services Region, either LONG_ID (17 characters), or SHORT_ID (8 characters).

Starting in October, 2021, you will receive an error when setting the account preference to SHORT_ID. Contact Amazon Web Services support if you receive an error and need to use short IDs for file system and mount target resources.

putAccountPreferencesResponse_resourceIdPreference :: Lens' PutAccountPreferencesResponse (Maybe ResourceIdPreference)

Undocumented member.

putAccountPreferencesResponse_httpStatus :: Lens' PutAccountPreferencesResponse Int

The response's http status code.

DescribeMountTargetSecurityGroups

describeMountTargetSecurityGroups_mountTargetId :: Lens' DescribeMountTargetSecurityGroups Text

The ID of the mount target whose security groups you want to retrieve.

describeMountTargetSecurityGroupsResponse_httpStatus :: Lens' DescribeMountTargetSecurityGroupsResponse Int

The response's http status code.

describeMountTargetSecurityGroupsResponse_securityGroups :: Lens' DescribeMountTargetSecurityGroupsResponse [Text]

An array of security groups.

DescribeAccessPoints

describeAccessPoints_accessPointId :: Lens' DescribeAccessPoints (Maybe Text)

(Optional) Specifies an EFS access point to describe in the response; mutually exclusive with FileSystemId.

describeAccessPoints_fileSystemId :: Lens' DescribeAccessPoints (Maybe Text)

(Optional) If you provide a FileSystemId, EFS returns all access points for that file system; mutually exclusive with AccessPointId.

describeAccessPoints_nextToken :: Lens' DescribeAccessPoints (Maybe Text)

NextToken is present if the response is paginated. You can use NextMarker in the subsequent request to fetch the next page of access point descriptions.

describeAccessPoints_maxResults :: Lens' DescribeAccessPoints (Maybe Natural)

(Optional) When retrieving all access points for a file system, you can optionally specify the MaxItems parameter to limit the number of objects returned in a response. The default value is 100.

describeAccessPointsResponse_accessPoints :: Lens' DescribeAccessPointsResponse (Maybe [AccessPointDescription])

An array of access point descriptions.

describeAccessPointsResponse_nextToken :: Lens' DescribeAccessPointsResponse (Maybe Text)

Present if there are more access points than returned in the response. You can use the NextMarker in the subsequent request to fetch the additional descriptions.

describeAccessPointsResponse_httpStatus :: Lens' DescribeAccessPointsResponse Int

The response's http status code.

ModifyMountTargetSecurityGroups

modifyMountTargetSecurityGroups_securityGroups :: Lens' ModifyMountTargetSecurityGroups (Maybe [Text])

An array of up to five VPC security group IDs.

modifyMountTargetSecurityGroups_mountTargetId :: Lens' ModifyMountTargetSecurityGroups Text

The ID of the mount target whose security groups you want to modify.

CreateFileSystem

createFileSystem_provisionedThroughputInMibps :: Lens' CreateFileSystem (Maybe Double)

The throughput, measured in MiB/s, that you want to provision for a file system that you're creating. Valid values are 1-1024. Required if ThroughputMode is set to provisioned. The upper limit for throughput is 1024 MiB/s. To increase this limit, contact Amazon Web Services Support. For more information, see Amazon EFS quotas that you can increase in the Amazon EFS User Guide.

createFileSystem_availabilityZoneName :: Lens' CreateFileSystem (Maybe Text)

Used to create a file system that uses One Zone storage classes. It specifies the Amazon Web Services Availability Zone in which to create the file system. Use the format us-east-1a to specify the Availability Zone. For more information about One Zone storage classes, see Using EFS storage classes in the Amazon EFS User Guide.

One Zone storage classes are not available in all Availability Zones in Amazon Web Services Regions where Amazon EFS is available.

createFileSystem_performanceMode :: Lens' CreateFileSystem (Maybe PerformanceMode)

The performance mode of the file system. We recommend generalPurpose performance mode for most file systems. File systems using the maxIO performance mode can scale to higher levels of aggregate throughput and operations per second with a tradeoff of slightly higher latencies for most file operations. The performance mode can't be changed after the file system has been created.

The maxIO mode is not supported on file systems using One Zone storage classes.

createFileSystem_backup :: Lens' CreateFileSystem (Maybe Bool)

Specifies whether automatic backups are enabled on the file system that you are creating. Set the value to true to enable automatic backups. If you are creating a file system that uses One Zone storage classes, automatic backups are enabled by default. For more information, see Automatic backups in the Amazon EFS User Guide.

Default is false. However, if you specify an AvailabilityZoneName, the default is true.

Backup is not available in all Amazon Web Services Regionswhere Amazon EFS is available.

createFileSystem_encrypted :: Lens' CreateFileSystem (Maybe Bool)

A Boolean value that, if true, creates an encrypted file system. When creating an encrypted file system, you have the option of specifying CreateFileSystemRequest$KmsKeyId for an existing Key Management Service (KMS customer master key (CMK). If you don't specify a CMK, then the default CMK for Amazon EFS, /aws/elasticfilesystem, is used to protect the encrypted file system.

createFileSystem_throughputMode :: Lens' CreateFileSystem (Maybe ThroughputMode)

Specifies the throughput mode for the file system, either bursting or provisioned. If you set ThroughputMode to provisioned, you must also set a value for ProvisionedThroughputInMibps. After you create the file system, you can decrease your file system's throughput in Provisioned Throughput mode or change between the throughput modes, as long as it’s been more than 24 hours since the last decrease or throughput mode change. For more information, see Specifying throughput with provisioned mode in the Amazon EFS User Guide.

Default is bursting.

createFileSystem_kmsKeyId :: Lens' CreateFileSystem (Maybe Text)

The ID of the KMS CMK that you want to use to protect the encrypted file system. This parameter is only required if you want to use a non-default KMS key. If this parameter is not specified, the default CMK for Amazon EFS is used. This ID can be in one of the following formats:

Key ID - A unique identifier of the key, for example 1234abcd-12ab-34cd-56ef-1234567890ab.

ARN - An Amazon Resource Name (ARN) for the key, for example arn:aws:kms:us-west-2:111122223333:key/1234abcd-12ab-34cd-56ef-1234567890ab.

Key alias - A previously created display name for a key, for example alias/projectKey1.

Key alias ARN - An ARN for a key alias, for example arn:aws:kms:us-west-2:444455556666:alias/projectKey1.

If KmsKeyId is specified, the CreateFileSystemRequest$Encrypted parameter must be set to true.

EFS accepts only symmetric KMS keys. You cannot use asymmetric KMS keys with EFS file systems.

createFileSystem_tags :: Lens' CreateFileSystem (Maybe [Tag])

Use to create one or more tags associated with the file system. Each tag is a user-defined key-value pair. Name your file system on creation by including a "Key":"Name","Value":"{value}" key-value pair. Each key must be unique. For more information, see Tagging Amazon Web Services resources in the Amazon Web Services General Reference Guide.

createFileSystem_creationToken :: Lens' CreateFileSystem Text

A string of up to 64 ASCII characters. Amazon EFS uses this to ensure idempotent creation.

fileSystemDescription_availabilityZoneId :: Lens' FileSystemDescription (Maybe Text)

The unique and consistent identifier of the Availability Zone in which the file system's One Zone storage classes exist. For example, use1-az1 is an Availability Zone ID for the us-east-1 Amazon Web Services Region, and it has the same location in every Amazon Web Services account.

fileSystemDescription_provisionedThroughputInMibps :: Lens' FileSystemDescription (Maybe Double)

The amount of provisioned throughput, measured in MiB/s, for the file system. Valid for file systems using ThroughputMode set to provisioned.

fileSystemDescription_availabilityZoneName :: Lens' FileSystemDescription (Maybe Text)

Describes the Amazon Web Services Availability Zone in which the file system is located, and is valid only for file systems using One Zone storage classes. For more information, see Using EFS storage classes in the Amazon EFS User Guide.

fileSystemDescription_fileSystemArn :: Lens' FileSystemDescription (Maybe Text)

The Amazon Resource Name (ARN) for the EFS file system, in the format arn:aws:elasticfilesystem:region:account-id:file-system/file-system-id . Example with sample data: arn:aws:elasticfilesystem:us-west-2:1111333322228888:file-system/fs-01234567

fileSystemDescription_encrypted :: Lens' FileSystemDescription (Maybe Bool)

A Boolean value that, if true, indicates that the file system is encrypted.

fileSystemDescription_throughputMode :: Lens' FileSystemDescription (Maybe ThroughputMode)

Displays the file system's throughput mode. For more information, see Throughput modes in the Amazon EFS User Guide.

fileSystemDescription_kmsKeyId :: Lens' FileSystemDescription (Maybe Text)

The ID of an Key Management Service customer master key (CMK) that was used to protect the encrypted file system.

fileSystemDescription_name :: Lens' FileSystemDescription (Maybe Text)

You can add tags to a file system, including a Name tag. For more information, see CreateFileSystem. If the file system has a Name tag, Amazon EFS returns the value in this field.

fileSystemDescription_ownerId :: Lens' FileSystemDescription Text

The Amazon Web Services account that created the file system. If the file system was created by an IAM user, the parent account to which the user belongs is the owner.

fileSystemDescription_creationToken :: Lens' FileSystemDescription Text

The opaque string specified in the request.

fileSystemDescription_fileSystemId :: Lens' FileSystemDescription Text

The ID of the file system, assigned by Amazon EFS.

fileSystemDescription_creationTime :: Lens' FileSystemDescription UTCTime

The time that the file system was created, in seconds (since 1970-01-01T00:00:00Z).

fileSystemDescription_lifeCycleState :: Lens' FileSystemDescription LifeCycleState

The lifecycle phase of the file system.

fileSystemDescription_numberOfMountTargets :: Lens' FileSystemDescription Natural

The current number of mount targets that the file system has. For more information, see CreateMountTarget.

fileSystemDescription_sizeInBytes :: Lens' FileSystemDescription FileSystemSize

The latest known metered size (in bytes) of data stored in the file system, in its Value field, and the time at which that size was determined in its Timestamp field. The Timestamp value is the integer number of seconds since 1970-01-01T00:00:00Z. The SizeInBytes value doesn't represent the size of a consistent snapshot of the file system, but it is eventually consistent when there are no writes to the file system. That is, SizeInBytes represents actual size only if the file system is not modified for a period longer than a couple of hours. Otherwise, the value is not the exact size that the file system was at any point in time.

fileSystemDescription_performanceMode :: Lens' FileSystemDescription PerformanceMode

The performance mode of the file system.

fileSystemDescription_tags :: Lens' FileSystemDescription [Tag]

The tags associated with the file system, presented as an array of Tag objects.

PutLifecycleConfiguration

putLifecycleConfiguration_fileSystemId :: Lens' PutLifecycleConfiguration Text

The ID of the file system for which you are creating the LifecycleConfiguration object (String).

putLifecycleConfiguration_lifecyclePolicies :: Lens' PutLifecycleConfiguration [LifecyclePolicy]

An array of LifecyclePolicy objects that define the file system's LifecycleConfiguration object. A LifecycleConfiguration object informs EFS lifecycle management and intelligent tiering of the following:

When to move files in the file system from primary storage to the IA storage class.

When to move files that are in IA storage to primary storage.

When using the put-lifecycle-configuration CLI command or the PutLifecycleConfiguration API action, Amazon EFS requires that each LifecyclePolicy object have only a single transition. This means that in a request body, LifecyclePolicies needs to be structured as an array of LifecyclePolicy objects, one object for each transition, TransitionToIA, TransitionToPrimaryStorageClass. See the example requests in the following section for more information.

lifecycleConfigurationDescription_lifecyclePolicies :: Lens' LifecycleConfigurationDescription (Maybe [LifecyclePolicy])

An array of lifecycle management policies. EFS supports a maximum of one policy per file system.

PutBackupPolicy

putBackupPolicy_fileSystemId :: Lens' PutBackupPolicy Text

Specifies which EFS file system to update the backup policy for.

putBackupPolicy_backupPolicy :: Lens' PutBackupPolicy BackupPolicy

The backup policy included in the PutBackupPolicy request.

backupPolicyDescription_backupPolicy :: Lens' BackupPolicyDescription (Maybe BackupPolicy)

Describes the file system's backup policy, indicating whether automatic backups are turned on or off..

DeleteFileSystem

deleteFileSystem_fileSystemId :: Lens' DeleteFileSystem Text

The ID of the file system you want to delete.

UpdateFileSystem

updateFileSystem_provisionedThroughputInMibps :: Lens' UpdateFileSystem (Maybe Double)

(Optional) Sets the amount of provisioned throughput, in MiB/s, for the file system. Valid values are 1-1024. If you are changing the throughput mode to provisioned, you must also provide the amount of provisioned throughput. Required if ThroughputMode is changed to provisioned on update.

updateFileSystem_throughputMode :: Lens' UpdateFileSystem (Maybe ThroughputMode)

(Optional) Updates the file system's throughput mode. If you're not updating your throughput mode, you don't need to provide this value in your request. If you are changing the ThroughputMode to provisioned, you must also set a value for ProvisionedThroughputInMibps.

updateFileSystem_fileSystemId :: Lens' UpdateFileSystem Text

The ID of the file system that you want to update.

fileSystemDescription_availabilityZoneId :: Lens' FileSystemDescription (Maybe Text)

The unique and consistent identifier of the Availability Zone in which the file system's One Zone storage classes exist. For example, use1-az1 is an Availability Zone ID for the us-east-1 Amazon Web Services Region, and it has the same location in every Amazon Web Services account.

fileSystemDescription_provisionedThroughputInMibps :: Lens' FileSystemDescription (Maybe Double)

The amount of provisioned throughput, measured in MiB/s, for the file system. Valid for file systems using ThroughputMode set to provisioned.

fileSystemDescription_availabilityZoneName :: Lens' FileSystemDescription (Maybe Text)

Describes the Amazon Web Services Availability Zone in which the file system is located, and is valid only for file systems using One Zone storage classes. For more information, see Using EFS storage classes in the Amazon EFS User Guide.

fileSystemDescription_fileSystemArn :: Lens' FileSystemDescription (Maybe Text)

The Amazon Resource Name (ARN) for the EFS file system, in the format arn:aws:elasticfilesystem:region:account-id:file-system/file-system-id . Example with sample data: arn:aws:elasticfilesystem:us-west-2:1111333322228888:file-system/fs-01234567

fileSystemDescription_encrypted :: Lens' FileSystemDescription (Maybe Bool)

A Boolean value that, if true, indicates that the file system is encrypted.

fileSystemDescription_throughputMode :: Lens' FileSystemDescription (Maybe ThroughputMode)

Displays the file system's throughput mode. For more information, see Throughput modes in the Amazon EFS User Guide.

fileSystemDescription_kmsKeyId :: Lens' FileSystemDescription (Maybe Text)

The ID of an Key Management Service customer master key (CMK) that was used to protect the encrypted file system.

fileSystemDescription_name :: Lens' FileSystemDescription (Maybe Text)

You can add tags to a file system, including a Name tag. For more information, see CreateFileSystem. If the file system has a Name tag, Amazon EFS returns the value in this field.

fileSystemDescription_ownerId :: Lens' FileSystemDescription Text

The Amazon Web Services account that created the file system. If the file system was created by an IAM user, the parent account to which the user belongs is the owner.

fileSystemDescription_creationToken :: Lens' FileSystemDescription Text

The opaque string specified in the request.

fileSystemDescription_fileSystemId :: Lens' FileSystemDescription Text

The ID of the file system, assigned by Amazon EFS.

fileSystemDescription_creationTime :: Lens' FileSystemDescription UTCTime

The time that the file system was created, in seconds (since 1970-01-01T00:00:00Z).

fileSystemDescription_lifeCycleState :: Lens' FileSystemDescription LifeCycleState

The lifecycle phase of the file system.

fileSystemDescription_numberOfMountTargets :: Lens' FileSystemDescription Natural

The current number of mount targets that the file system has. For more information, see CreateMountTarget.

fileSystemDescription_sizeInBytes :: Lens' FileSystemDescription FileSystemSize

The latest known metered size (in bytes) of data stored in the file system, in its Value field, and the time at which that size was determined in its Timestamp field. The Timestamp value is the integer number of seconds since 1970-01-01T00:00:00Z. The SizeInBytes value doesn't represent the size of a consistent snapshot of the file system, but it is eventually consistent when there are no writes to the file system. That is, SizeInBytes represents actual size only if the file system is not modified for a period longer than a couple of hours. Otherwise, the value is not the exact size that the file system was at any point in time.

fileSystemDescription_performanceMode :: Lens' FileSystemDescription PerformanceMode

The performance mode of the file system.

fileSystemDescription_tags :: Lens' FileSystemDescription [Tag]

The tags associated with the file system, presented as an array of Tag objects.

CreateMountTarget

createMountTarget_ipAddress :: Lens' CreateMountTarget (Maybe Text)

Valid IPv4 address within the address range of the specified subnet.

createMountTarget_securityGroups :: Lens' CreateMountTarget (Maybe [Text])

Up to five VPC security group IDs, of the form sg-xxxxxxxx. These must be for the same VPC as subnet specified.

createMountTarget_fileSystemId :: Lens' CreateMountTarget Text

The ID of the file system for which to create the mount target.

createMountTarget_subnetId :: Lens' CreateMountTarget Text

The ID of the subnet to add the mount target in. For file systems that use One Zone storage classes, use the subnet that is associated with the file system's Availability Zone.

mountTargetDescription_ipAddress :: Lens' MountTargetDescription (Maybe Text)

Address at which the file system can be mounted by using the mount target.

mountTargetDescription_availabilityZoneId :: Lens' MountTargetDescription (Maybe Text)

The unique and consistent identifier of the Availability Zone that the mount target resides in. For example, use1-az1 is an AZ ID for the us-east-1 Region and it has the same location in every Amazon Web Services account.

mountTargetDescription_vpcId :: Lens' MountTargetDescription (Maybe Text)

The virtual private cloud (VPC) ID that the mount target is configured in.

mountTargetDescription_availabilityZoneName :: Lens' MountTargetDescription (Maybe Text)

The name of the Availability Zone in which the mount target is located. Availability Zones are independently mapped to names for each Amazon Web Services account. For example, the Availability Zone us-east-1a for your Amazon Web Services account might not be the same location as us-east-1a for another Amazon Web Services account.

mountTargetDescription_networkInterfaceId :: Lens' MountTargetDescription (Maybe Text)

The ID of the network interface that Amazon EFS created when it created the mount target.

mountTargetDescription_ownerId :: Lens' MountTargetDescription (Maybe Text)

Amazon Web Services account ID that owns the resource.

mountTargetDescription_mountTargetId :: Lens' MountTargetDescription Text

System-assigned mount target ID.

mountTargetDescription_fileSystemId :: Lens' MountTargetDescription Text

The ID of the file system for which the mount target is intended.

mountTargetDescription_subnetId :: Lens' MountTargetDescription Text

The ID of the mount target's subnet.

mountTargetDescription_lifeCycleState :: Lens' MountTargetDescription LifeCycleState

Lifecycle state of the mount target.

TagResource

tagResource_resourceId :: Lens' TagResource Text

The ID specifying the EFS resource that you want to create a tag for.

tagResource_tags :: Lens' TagResource [Tag]

An array of Tag objects to add. Each Tag object is a key-value pair.

DescribeBackupPolicy

describeBackupPolicy_fileSystemId :: Lens' DescribeBackupPolicy Text

Specifies which EFS file system to retrieve the BackupPolicy for.

backupPolicyDescription_backupPolicy :: Lens' BackupPolicyDescription (Maybe BackupPolicy)

Describes the file system's backup policy, indicating whether automatic backups are turned on or off..

DescribeLifecycleConfiguration

describeLifecycleConfiguration_fileSystemId :: Lens' DescribeLifecycleConfiguration Text

The ID of the file system whose LifecycleConfiguration object you want to retrieve (String).

lifecycleConfigurationDescription_lifecyclePolicies :: Lens' LifecycleConfigurationDescription (Maybe [LifecyclePolicy])

An array of lifecycle management policies. EFS supports a maximum of one policy per file system.

UntagResource

untagResource_resourceId :: Lens' UntagResource Text

Specifies the EFS resource that you want to remove tags from.

untagResource_tagKeys :: Lens' UntagResource (NonEmpty Text)

The keys of the key-value tag pairs that you want to remove from the specified EFS resource.

DescribeFileSystemPolicy

describeFileSystemPolicy_fileSystemId :: Lens' DescribeFileSystemPolicy Text

Specifies which EFS file system to retrieve the FileSystemPolicy for.

fileSystemPolicyDescription_fileSystemId :: Lens' FileSystemPolicyDescription (Maybe Text)

Specifies the EFS file system to which the FileSystemPolicy applies.

fileSystemPolicyDescription_policy :: Lens' FileSystemPolicyDescription (Maybe Text)

The JSON formatted FileSystemPolicy for the EFS file system.

DeleteAccessPoint

deleteAccessPoint_accessPointId :: Lens' DeleteAccessPoint Text

The ID of the access point that you want to delete.

Types

AccessPointDescription

accessPointDescription_posixUser :: Lens' AccessPointDescription (Maybe PosixUser)

The full POSIX identity, including the user ID, group ID, and secondary group IDs on the access point that is used for all file operations by NFS clients using the access point.

accessPointDescription_rootDirectory :: Lens' AccessPointDescription (Maybe RootDirectory)

The directory on the Amazon EFS file system that the access point exposes as the root directory to NFS clients using the access point.

accessPointDescription_clientToken :: Lens' AccessPointDescription (Maybe Text)

The opaque string specified in the request to ensure idempotent creation.

accessPointDescription_accessPointId :: Lens' AccessPointDescription (Maybe Text)

The ID of the access point, assigned by Amazon EFS.

accessPointDescription_fileSystemId :: Lens' AccessPointDescription (Maybe Text)

The ID of the EFS file system that the access point applies to.

accessPointDescription_ownerId :: Lens' AccessPointDescription (Maybe Text)

Identified the Amazon Web Services account that owns the access point resource.

accessPointDescription_name :: Lens' AccessPointDescription (Maybe Text)

The name of the access point. This is the value of the Name tag.

accessPointDescription_accessPointArn :: Lens' AccessPointDescription (Maybe Text)

The unique Amazon Resource Name (ARN) associated with the access point.

accessPointDescription_lifeCycleState :: Lens' AccessPointDescription (Maybe LifeCycleState)

Identifies the lifecycle phase of the access point.

accessPointDescription_tags :: Lens' AccessPointDescription (Maybe [Tag])

The tags associated with the access point, presented as an array of Tag objects.

BackupPolicy

backupPolicy_status :: Lens' BackupPolicy BackupStatus

Describes the status of the file system's backup policy.

ENABLED - EFS is automatically backing up the file system.

ENABLING - EFS is turning on automatic backups for the file system.

DISABLED - automatic back ups are turned off for the file system.

DISABLING - EFS is turning off automatic backups for the file system.

BackupPolicyDescription

backupPolicyDescription_backupPolicy :: Lens' BackupPolicyDescription (Maybe BackupPolicy)

Describes the file system's backup policy, indicating whether automatic backups are turned on or off..

CreationInfo

creationInfo_ownerUid :: Lens' CreationInfo Natural

Specifies the POSIX user ID to apply to the RootDirectory. Accepts values from 0 to 2^32 (4294967295).

creationInfo_ownerGid :: Lens' CreationInfo Natural

Specifies the POSIX group ID to apply to the RootDirectory. Accepts values from 0 to 2^32 (4294967295).

creationInfo_permissions :: Lens' CreationInfo Text

Specifies the POSIX permissions to apply to the RootDirectory, in the format of an octal number representing the file's mode bits.

FileSystemDescription

fileSystemDescription_availabilityZoneId :: Lens' FileSystemDescription (Maybe Text)

The unique and consistent identifier of the Availability Zone in which the file system's One Zone storage classes exist. For example, use1-az1 is an Availability Zone ID for the us-east-1 Amazon Web Services Region, and it has the same location in every Amazon Web Services account.

fileSystemDescription_provisionedThroughputInMibps :: Lens' FileSystemDescription (Maybe Double)

The amount of provisioned throughput, measured in MiB/s, for the file system. Valid for file systems using ThroughputMode set to provisioned.

fileSystemDescription_availabilityZoneName :: Lens' FileSystemDescription (Maybe Text)

Describes the Amazon Web Services Availability Zone in which the file system is located, and is valid only for file systems using One Zone storage classes. For more information, see Using EFS storage classes in the Amazon EFS User Guide.

fileSystemDescription_fileSystemArn :: Lens' FileSystemDescription (Maybe Text)

The Amazon Resource Name (ARN) for the EFS file system, in the format arn:aws:elasticfilesystem:region:account-id:file-system/file-system-id . Example with sample data: arn:aws:elasticfilesystem:us-west-2:1111333322228888:file-system/fs-01234567

fileSystemDescription_encrypted :: Lens' FileSystemDescription (Maybe Bool)

A Boolean value that, if true, indicates that the file system is encrypted.

fileSystemDescription_throughputMode :: Lens' FileSystemDescription (Maybe ThroughputMode)

Displays the file system's throughput mode. For more information, see Throughput modes in the Amazon EFS User Guide.

fileSystemDescription_kmsKeyId :: Lens' FileSystemDescription (Maybe Text)

The ID of an Key Management Service customer master key (CMK) that was used to protect the encrypted file system.

fileSystemDescription_name :: Lens' FileSystemDescription (Maybe Text)

You can add tags to a file system, including a Name tag. For more information, see CreateFileSystem. If the file system has a Name tag, Amazon EFS returns the value in this field.

fileSystemDescription_ownerId :: Lens' FileSystemDescription Text

The Amazon Web Services account that created the file system. If the file system was created by an IAM user, the parent account to which the user belongs is the owner.

fileSystemDescription_creationToken :: Lens' FileSystemDescription Text

The opaque string specified in the request.

fileSystemDescription_fileSystemId :: Lens' FileSystemDescription Text

The ID of the file system, assigned by Amazon EFS.

fileSystemDescription_creationTime :: Lens' FileSystemDescription UTCTime

The time that the file system was created, in seconds (since 1970-01-01T00:00:00Z).

fileSystemDescription_lifeCycleState :: Lens' FileSystemDescription LifeCycleState

The lifecycle phase of the file system.

fileSystemDescription_numberOfMountTargets :: Lens' FileSystemDescription Natural

The current number of mount targets that the file system has. For more information, see CreateMountTarget.

fileSystemDescription_sizeInBytes :: Lens' FileSystemDescription FileSystemSize

The latest known metered size (in bytes) of data stored in the file system, in its Value field, and the time at which that size was determined in its Timestamp field. The Timestamp value is the integer number of seconds since 1970-01-01T00:00:00Z. The SizeInBytes value doesn't represent the size of a consistent snapshot of the file system, but it is eventually consistent when there are no writes to the file system. That is, SizeInBytes represents actual size only if the file system is not modified for a period longer than a couple of hours. Otherwise, the value is not the exact size that the file system was at any point in time.

fileSystemDescription_performanceMode :: Lens' FileSystemDescription PerformanceMode

The performance mode of the file system.

fileSystemDescription_tags :: Lens' FileSystemDescription [Tag]

The tags associated with the file system, presented as an array of Tag objects.

FileSystemPolicyDescription

fileSystemPolicyDescription_fileSystemId :: Lens' FileSystemPolicyDescription (Maybe Text)

Specifies the EFS file system to which the FileSystemPolicy applies.

fileSystemPolicyDescription_policy :: Lens' FileSystemPolicyDescription (Maybe Text)

The JSON formatted FileSystemPolicy for the EFS file system.

FileSystemSize

fileSystemSize_valueInIA :: Lens' FileSystemSize (Maybe Natural)

The latest known metered size (in bytes) of data stored in the Infrequent Access storage class.

fileSystemSize_valueInStandard :: Lens' FileSystemSize (Maybe Natural)

The latest known metered size (in bytes) of data stored in the Standard storage class.

fileSystemSize_timestamp :: Lens' FileSystemSize (Maybe UTCTime)

The time at which the size of data, returned in the Value field, was determined. The value is the integer number of seconds since 1970-01-01T00:00:00Z.

fileSystemSize_value :: Lens' FileSystemSize Natural

The latest known metered size (in bytes) of data stored in the file system.

LifecycleConfigurationDescription

lifecycleConfigurationDescription_lifecyclePolicies :: Lens' LifecycleConfigurationDescription (Maybe [LifecyclePolicy])

An array of lifecycle management policies. EFS supports a maximum of one policy per file system.

LifecyclePolicy

lifecyclePolicy_transitionToIA :: Lens' LifecyclePolicy (Maybe TransitionToIARules)

Describes the period of time that a file is not accessed, after which it transitions to IA storage. Metadata operations such as listing the contents of a directory don't count as file access events.

lifecyclePolicy_transitionToPrimaryStorageClass :: Lens' LifecyclePolicy (Maybe TransitionToPrimaryStorageClassRules)

Describes when to transition a file from IA storage to primary storage. Metadata operations such as listing the contents of a directory don't count as file access events.

MountTargetDescription

mountTargetDescription_ipAddress :: Lens' MountTargetDescription (Maybe Text)

Address at which the file system can be mounted by using the mount target.

mountTargetDescription_availabilityZoneId :: Lens' MountTargetDescription (Maybe Text)

The unique and consistent identifier of the Availability Zone that the mount target resides in. For example, use1-az1 is an AZ ID for the us-east-1 Region and it has the same location in every Amazon Web Services account.

mountTargetDescription_vpcId :: Lens' MountTargetDescription (Maybe Text)

The virtual private cloud (VPC) ID that the mount target is configured in.

mountTargetDescription_availabilityZoneName :: Lens' MountTargetDescription (Maybe Text)

The name of the Availability Zone in which the mount target is located. Availability Zones are independently mapped to names for each Amazon Web Services account. For example, the Availability Zone us-east-1a for your Amazon Web Services account might not be the same location as us-east-1a for another Amazon Web Services account.

mountTargetDescription_networkInterfaceId :: Lens' MountTargetDescription (Maybe Text)

The ID of the network interface that Amazon EFS created when it created the mount target.

mountTargetDescription_ownerId :: Lens' MountTargetDescription (Maybe Text)

Amazon Web Services account ID that owns the resource.

mountTargetDescription_mountTargetId :: Lens' MountTargetDescription Text

System-assigned mount target ID.

mountTargetDescription_fileSystemId :: Lens' MountTargetDescription Text

The ID of the file system for which the mount target is intended.

mountTargetDescription_subnetId :: Lens' MountTargetDescription Text

The ID of the mount target's subnet.

mountTargetDescription_lifeCycleState :: Lens' MountTargetDescription LifeCycleState

Lifecycle state of the mount target.

PosixUser

posixUser_secondaryGids :: Lens' PosixUser (Maybe [Natural])

Secondary POSIX group IDs used for all file system operations using this access point.

posixUser_uid :: Lens' PosixUser Natural

The POSIX user ID used for all file system operations using this access point.

posixUser_gid :: Lens' PosixUser Natural

The POSIX group ID used for all file system operations using this access point.

ResourceIdPreference

resourceIdPreference_resources :: Lens' ResourceIdPreference (Maybe [Resource])

Identifies the Amazon EFS resources to which the ID preference setting applies, FILE_SYSTEM and MOUNT_TARGET.

resourceIdPreference_resourceIdType :: Lens' ResourceIdPreference (Maybe ResourceIdType)

Identifies the EFS resource ID preference, either LONG_ID (17 characters) or SHORT_ID (8 characters).

RootDirectory

rootDirectory_creationInfo :: Lens' RootDirectory (Maybe CreationInfo)

(Optional) Specifies the POSIX IDs and permissions to apply to the access point's RootDirectory. If the RootDirectory > Path specified does not exist, EFS creates the root directory using the CreationInfo settings when a client connects to an access point. When specifying the CreationInfo, you must provide values for all properties.

If you do not provide CreationInfo and the specified RootDirectory > Path does not exist, attempts to mount the file system using the access point will fail.

rootDirectory_path :: Lens' RootDirectory (Maybe Text)

Specifies the path on the EFS file system to expose as the root directory to NFS clients using the access point to access the EFS file system. A path can have up to four subdirectories. If the specified path does not exist, you are required to provide the CreationInfo.

Tag

tag_key :: Lens' Tag Text

The tag key (String). The key can't start with aws:.

tag_value :: Lens' Tag Text

The value of the tag key.