Amazonka.EFS.Types.AccessPointDescription

Description

Documentation

data AccessPointDescription

Provides a description of an EFS file system access point.

See: newAccessPointDescription smart constructor.

Constructors

AccessPointDescription'

Fields

posixUser :: Maybe PosixUser

The full POSIX identity, including the user ID, group ID, and secondary group IDs on the access point that is used for all file operations by NFS clients using the access point.

rootDirectory :: Maybe RootDirectory

The directory on the Amazon EFS file system that the access point exposes as the root directory to NFS clients using the access point.

clientToken :: Maybe Text

The opaque string specified in the request to ensure idempotent creation.

accessPointId :: Maybe Text

The ID of the access point, assigned by Amazon EFS.

fileSystemId :: Maybe Text

The ID of the EFS file system that the access point applies to.

ownerId :: Maybe Text

Identified the Amazon Web Services account that owns the access point resource.

name :: Maybe Text

The name of the access point. This is the value of the Name tag.

accessPointArn :: Maybe Text

The unique Amazon Resource Name (ARN) associated with the access point.

lifeCycleState :: Maybe LifeCycleState

Identifies the lifecycle phase of the access point.

tags :: Maybe [Tag]

The tags associated with the access point, presented as an array of Tag objects.

newAccessPointDescription :: AccessPointDescription

Create a value of AccessPointDescription with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:posixUser:AccessPointDescription', accessPointDescription_posixUser - The full POSIX identity, including the user ID, group ID, and secondary group IDs on the access point that is used for all file operations by NFS clients using the access point.

$sel:rootDirectory:AccessPointDescription', accessPointDescription_rootDirectory - The directory on the Amazon EFS file system that the access point exposes as the root directory to NFS clients using the access point.

$sel:clientToken:AccessPointDescription', accessPointDescription_clientToken - The opaque string specified in the request to ensure idempotent creation.

$sel:accessPointId:AccessPointDescription', accessPointDescription_accessPointId - The ID of the access point, assigned by Amazon EFS.

$sel:fileSystemId:AccessPointDescription', accessPointDescription_fileSystemId - The ID of the EFS file system that the access point applies to.

$sel:ownerId:AccessPointDescription', accessPointDescription_ownerId - Identified the Amazon Web Services account that owns the access point resource.

$sel:name:AccessPointDescription', accessPointDescription_name - The name of the access point. This is the value of the Name tag.

$sel:accessPointArn:AccessPointDescription', accessPointDescription_accessPointArn - The unique Amazon Resource Name (ARN) associated with the access point.

$sel:lifeCycleState:AccessPointDescription', accessPointDescription_lifeCycleState - Identifies the lifecycle phase of the access point.

$sel:tags:AccessPointDescription', accessPointDescription_tags - The tags associated with the access point, presented as an array of Tag objects.

accessPointDescription_posixUser :: Lens' AccessPointDescription (Maybe PosixUser)

The full POSIX identity, including the user ID, group ID, and secondary group IDs on the access point that is used for all file operations by NFS clients using the access point.

accessPointDescription_rootDirectory :: Lens' AccessPointDescription (Maybe RootDirectory)

The directory on the Amazon EFS file system that the access point exposes as the root directory to NFS clients using the access point.

accessPointDescription_clientToken :: Lens' AccessPointDescription (Maybe Text)

The opaque string specified in the request to ensure idempotent creation.

accessPointDescription_accessPointId :: Lens' AccessPointDescription (Maybe Text)

The ID of the access point, assigned by Amazon EFS.

accessPointDescription_fileSystemId :: Lens' AccessPointDescription (Maybe Text)

The ID of the EFS file system that the access point applies to.

accessPointDescription_ownerId :: Lens' AccessPointDescription (Maybe Text)

Identified the Amazon Web Services account that owns the access point resource.

accessPointDescription_name :: Lens' AccessPointDescription (Maybe Text)

The name of the access point. This is the value of the Name tag.

accessPointDescription_accessPointArn :: Lens' AccessPointDescription (Maybe Text)

The unique Amazon Resource Name (ARN) associated with the access point.

accessPointDescription_lifeCycleState :: Lens' AccessPointDescription (Maybe LifeCycleState)

Identifies the lifecycle phase of the access point.

accessPointDescription_tags :: Lens' AccessPointDescription (Maybe [Tag])

The tags associated with the access point, presented as an array of Tag objects.