Amazonka.EKS.DeleteAddon

Description

Delete an Amazon EKS add-on.

When you remove the add-on, it will also be deleted from the cluster. You can always manually start an add-on on the cluster using the Kubernetes API.

Creating a Request

data DeleteAddon

See: newDeleteAddon smart constructor.

Constructors

DeleteAddon'

Fields

preserve :: Maybe Bool

Specifying this option preserves the add-on software on your cluster but Amazon EKS stops managing any settings for the add-on. If an IAM account is associated with the add-on, it is not removed.

clusterName :: Text

The name of the cluster to delete the add-on from.

addonName :: Text

The name of the add-on. The name must match one of the names returned by ListAddons .

newDeleteAddon

Arguments

:: Text
$sel:clusterName:DeleteAddon'

-> Text
$sel:addonName:DeleteAddon'

-> DeleteAddon

Create a value of DeleteAddon with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:preserve:DeleteAddon', deleteAddon_preserve - Specifying this option preserves the add-on software on your cluster but Amazon EKS stops managing any settings for the add-on. If an IAM account is associated with the add-on, it is not removed.

$sel:clusterName:DeleteAddon', deleteAddon_clusterName - The name of the cluster to delete the add-on from.

$sel:addonName:DeleteAddon', deleteAddon_addonName - The name of the add-on. The name must match one of the names returned by ListAddons .

Request Lenses

deleteAddon_preserve :: Lens' DeleteAddon (Maybe Bool)

Specifying this option preserves the add-on software on your cluster but Amazon EKS stops managing any settings for the add-on. If an IAM account is associated with the add-on, it is not removed.

deleteAddon_clusterName :: Lens' DeleteAddon Text

The name of the cluster to delete the add-on from.

deleteAddon_addonName :: Lens' DeleteAddon Text

The name of the add-on. The name must match one of the names returned by ListAddons .

Destructuring the Response

data DeleteAddonResponse

See: newDeleteAddonResponse smart constructor.

Constructors

DeleteAddonResponse'

Fields

addon :: Maybe Addon

httpStatus :: Int

The response's http status code.

newDeleteAddonResponse

Arguments

:: Int
$sel:httpStatus:DeleteAddonResponse'

-> DeleteAddonResponse

Create a value of DeleteAddonResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:addon:DeleteAddonResponse', deleteAddonResponse_addon - Undocumented member.

$sel:httpStatus:DeleteAddonResponse', deleteAddonResponse_httpStatus - The response's http status code.

Response Lenses

deleteAddonResponse_addon :: Lens' DeleteAddonResponse (Maybe Addon)

Undocumented member.

deleteAddonResponse_httpStatus :: Lens' DeleteAddonResponse Int

The response's http status code.