Amazonka.EKS.DescribeAddon

Description

Describes an Amazon EKS add-on.

Creating a Request

data DescribeAddon

See: newDescribeAddon smart constructor.

Constructors

DescribeAddon'

Fields

clusterName :: Text

The name of the cluster.

addonName :: Text

The name of the add-on. The name must match one of the names returned by ListAddons .

newDescribeAddon

Arguments

:: Text
$sel:clusterName:DescribeAddon'

-> Text
$sel:addonName:DescribeAddon'

-> DescribeAddon

Create a value of DescribeAddon with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:clusterName:DescribeAddon', describeAddon_clusterName - The name of the cluster.

$sel:addonName:DescribeAddon', describeAddon_addonName - The name of the add-on. The name must match one of the names returned by ListAddons .

Request Lenses

describeAddon_clusterName :: Lens' DescribeAddon Text

The name of the cluster.

describeAddon_addonName :: Lens' DescribeAddon Text

The name of the add-on. The name must match one of the names returned by ListAddons .

Destructuring the Response

data DescribeAddonResponse

See: newDescribeAddonResponse smart constructor.

Constructors

DescribeAddonResponse'

Fields

addon :: Maybe Addon

httpStatus :: Int

The response's http status code.

newDescribeAddonResponse

Arguments

:: Int
$sel:httpStatus:DescribeAddonResponse'

-> DescribeAddonResponse

Create a value of DescribeAddonResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:addon:DescribeAddonResponse', describeAddonResponse_addon - Undocumented member.

$sel:httpStatus:DescribeAddonResponse', describeAddonResponse_httpStatus - The response's http status code.

Response Lenses

describeAddonResponse_addon :: Lens' DescribeAddonResponse (Maybe Addon)

Undocumented member.

describeAddonResponse_httpStatus :: Lens' DescribeAddonResponse Int

The response's http status code.