Amazonka.EKS.ListIdentityProviderConfigs

Description

A list of identity provider configurations.

This operation returns paginated results.

Creating a Request

data ListIdentityProviderConfigs

See: newListIdentityProviderConfigs smart constructor.

Constructors

ListIdentityProviderConfigs'

Fields

nextToken :: Maybe Text

The nextToken value returned from a previous paginated IdentityProviderConfigsRequest where maxResults was used and the results exceeded the value of that parameter. Pagination continues from the end of the previous results that returned the nextToken value.

maxResults :: Maybe Natural

The maximum number of identity provider configurations returned by ListIdentityProviderConfigs in paginated output. When you use this parameter, ListIdentityProviderConfigs returns only maxResults results in a single page along with a nextToken response element. You can see the remaining results of the initial request by sending another ListIdentityProviderConfigs request with the returned nextToken value. This value can be between 1 and 100. If you don't use this parameter, ListIdentityProviderConfigs returns up to 100 results and a nextToken value, if applicable.

clusterName :: Text

The cluster name that you want to list identity provider configurations for.

newListIdentityProviderConfigs

Arguments

:: Text
$sel:clusterName:ListIdentityProviderConfigs'

-> ListIdentityProviderConfigs

Create a value of ListIdentityProviderConfigs with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:nextToken:ListIdentityProviderConfigs', listIdentityProviderConfigs_nextToken - The nextToken value returned from a previous paginated IdentityProviderConfigsRequest where maxResults was used and the results exceeded the value of that parameter. Pagination continues from the end of the previous results that returned the nextToken value.

$sel:maxResults:ListIdentityProviderConfigs', listIdentityProviderConfigs_maxResults - The maximum number of identity provider configurations returned by ListIdentityProviderConfigs in paginated output. When you use this parameter, ListIdentityProviderConfigs returns only maxResults results in a single page along with a nextToken response element. You can see the remaining results of the initial request by sending another ListIdentityProviderConfigs request with the returned nextToken value. This value can be between 1 and 100. If you don't use this parameter, ListIdentityProviderConfigs returns up to 100 results and a nextToken value, if applicable.

$sel:clusterName:ListIdentityProviderConfigs', listIdentityProviderConfigs_clusterName - The cluster name that you want to list identity provider configurations for.

Request Lenses

listIdentityProviderConfigs_nextToken :: Lens' ListIdentityProviderConfigs (Maybe Text)

The nextToken value returned from a previous paginated IdentityProviderConfigsRequest where maxResults was used and the results exceeded the value of that parameter. Pagination continues from the end of the previous results that returned the nextToken value.

listIdentityProviderConfigs_maxResults :: Lens' ListIdentityProviderConfigs (Maybe Natural)

The maximum number of identity provider configurations returned by ListIdentityProviderConfigs in paginated output. When you use this parameter, ListIdentityProviderConfigs returns only maxResults results in a single page along with a nextToken response element. You can see the remaining results of the initial request by sending another ListIdentityProviderConfigs request with the returned nextToken value. This value can be between 1 and 100. If you don't use this parameter, ListIdentityProviderConfigs returns up to 100 results and a nextToken value, if applicable.

listIdentityProviderConfigs_clusterName :: Lens' ListIdentityProviderConfigs Text

The cluster name that you want to list identity provider configurations for.

Destructuring the Response

data ListIdentityProviderConfigsResponse

See: newListIdentityProviderConfigsResponse smart constructor.

Constructors

ListIdentityProviderConfigsResponse'

Fields

identityProviderConfigs :: Maybe [IdentityProviderConfig]

The identity provider configurations for the cluster.

nextToken :: Maybe Text

The nextToken value returned from a previous paginated ListIdentityProviderConfigsResponse where maxResults was used and the results exceeded the value of that parameter. Pagination continues from the end of the previous results that returned the nextToken value.

httpStatus :: Int

The response's http status code.

newListIdentityProviderConfigsResponse

Arguments

:: Int
$sel:httpStatus:ListIdentityProviderConfigsResponse'

-> ListIdentityProviderConfigsResponse

Create a value of ListIdentityProviderConfigsResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:identityProviderConfigs:ListIdentityProviderConfigsResponse', listIdentityProviderConfigsResponse_identityProviderConfigs - The identity provider configurations for the cluster.

$sel:nextToken:ListIdentityProviderConfigs', listIdentityProviderConfigsResponse_nextToken - The nextToken value returned from a previous paginated ListIdentityProviderConfigsResponse where maxResults was used and the results exceeded the value of that parameter. Pagination continues from the end of the previous results that returned the nextToken value.

$sel:httpStatus:ListIdentityProviderConfigsResponse', listIdentityProviderConfigsResponse_httpStatus - The response's http status code.

Response Lenses

listIdentityProviderConfigsResponse_identityProviderConfigs :: Lens' ListIdentityProviderConfigsResponse (Maybe [IdentityProviderConfig])

The identity provider configurations for the cluster.

listIdentityProviderConfigsResponse_nextToken :: Lens' ListIdentityProviderConfigsResponse (Maybe Text)

The nextToken value returned from a previous paginated ListIdentityProviderConfigsResponse where maxResults was used and the results exceeded the value of that parameter. Pagination continues from the end of the previous results that returned the nextToken value.

listIdentityProviderConfigsResponse_httpStatus :: Lens' ListIdentityProviderConfigsResponse Int

The response's http status code.