Amazonka.EKS.UpdateAddon

Description

Updates an Amazon EKS add-on.

Creating a Request

data UpdateAddon

See: newUpdateAddon smart constructor.

Constructors

UpdateAddon'

Fields

addonVersion :: Maybe Text

The version of the add-on. The version must match one of the versions returned by DescribeAddonVersions .

serviceAccountRoleArn :: Maybe Text

The Amazon Resource Name (ARN) of an existing IAM role to bind to the add-on's service account. The role must be assigned the IAM permissions required by the add-on. If you don't specify an existing IAM role, then the add-on uses the permissions assigned to the node IAM role. For more information, see Amazon EKS node IAM role in the Amazon EKS User Guide.

To specify an existing IAM role, you must have an IAM OpenID Connect (OIDC) provider created for your cluster. For more information, see Enabling IAM roles for service accounts on your cluster in the Amazon EKS User Guide.

resolveConflicts :: Maybe ResolveConflicts

How to resolve parameter value conflicts when applying the new version of the add-on to the cluster.

clientRequestToken :: Maybe Text

Unique, case-sensitive identifier that you provide to ensure the idempotency of the request.

clusterName :: Text

The name of the cluster.

addonName :: Text

The name of the add-on. The name must match one of the names returned by ListAddons .

newUpdateAddon

Arguments

:: Text
$sel:clusterName:UpdateAddon'

-> Text
$sel:addonName:UpdateAddon'

-> UpdateAddon

Create a value of UpdateAddon with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:addonVersion:UpdateAddon', updateAddon_addonVersion - The version of the add-on. The version must match one of the versions returned by DescribeAddonVersions .

$sel:serviceAccountRoleArn:UpdateAddon', updateAddon_serviceAccountRoleArn - The Amazon Resource Name (ARN) of an existing IAM role to bind to the add-on's service account. The role must be assigned the IAM permissions required by the add-on. If you don't specify an existing IAM role, then the add-on uses the permissions assigned to the node IAM role. For more information, see Amazon EKS node IAM role in the Amazon EKS User Guide.

To specify an existing IAM role, you must have an IAM OpenID Connect (OIDC) provider created for your cluster. For more information, see Enabling IAM roles for service accounts on your cluster in the Amazon EKS User Guide.

$sel:resolveConflicts:UpdateAddon', updateAddon_resolveConflicts - How to resolve parameter value conflicts when applying the new version of the add-on to the cluster.

$sel:clientRequestToken:UpdateAddon', updateAddon_clientRequestToken - Unique, case-sensitive identifier that you provide to ensure the idempotency of the request.

$sel:clusterName:UpdateAddon', updateAddon_clusterName - The name of the cluster.

$sel:addonName:UpdateAddon', updateAddon_addonName - The name of the add-on. The name must match one of the names returned by ListAddons .

Request Lenses

updateAddon_addonVersion :: Lens' UpdateAddon (Maybe Text)

The version of the add-on. The version must match one of the versions returned by DescribeAddonVersions .

updateAddon_serviceAccountRoleArn :: Lens' UpdateAddon (Maybe Text)

The Amazon Resource Name (ARN) of an existing IAM role to bind to the add-on's service account. The role must be assigned the IAM permissions required by the add-on. If you don't specify an existing IAM role, then the add-on uses the permissions assigned to the node IAM role. For more information, see Amazon EKS node IAM role in the Amazon EKS User Guide.

To specify an existing IAM role, you must have an IAM OpenID Connect (OIDC) provider created for your cluster. For more information, see Enabling IAM roles for service accounts on your cluster in the Amazon EKS User Guide.

updateAddon_resolveConflicts :: Lens' UpdateAddon (Maybe ResolveConflicts)

How to resolve parameter value conflicts when applying the new version of the add-on to the cluster.

updateAddon_clientRequestToken :: Lens' UpdateAddon (Maybe Text)

Unique, case-sensitive identifier that you provide to ensure the idempotency of the request.

updateAddon_clusterName :: Lens' UpdateAddon Text

The name of the cluster.

updateAddon_addonName :: Lens' UpdateAddon Text

The name of the add-on. The name must match one of the names returned by ListAddons .

Destructuring the Response

data UpdateAddonResponse

See: newUpdateAddonResponse smart constructor.

Constructors

UpdateAddonResponse'

Fields

update :: Maybe Update

httpStatus :: Int

The response's http status code.

newUpdateAddonResponse

Arguments

:: Int
$sel:httpStatus:UpdateAddonResponse'

-> UpdateAddonResponse

Create a value of UpdateAddonResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:update:UpdateAddonResponse', updateAddonResponse_update - Undocumented member.

$sel:httpStatus:UpdateAddonResponse', updateAddonResponse_httpStatus - The response's http status code.

Response Lenses

updateAddonResponse_update :: Lens' UpdateAddonResponse (Maybe Update)

Undocumented member.

updateAddonResponse_httpStatus :: Lens' UpdateAddonResponse Int

The response's http status code.