Amazonka.ELB.CreateLBCookieStickinessPolicy

Description

Generates a stickiness policy with sticky session lifetimes controlled by the lifetime of the browser (user-agent) or a specified expiration period. This policy can be associated only with HTTP/HTTPS listeners.

When a load balancer implements this policy, the load balancer uses a special cookie to track the instance for each request. When the load balancer receives a request, it first checks to see if this cookie is present in the request. If so, the load balancer sends the request to the application server specified in the cookie. If not, the load balancer sends the request to a server that is chosen based on the existing load-balancing algorithm.

A cookie is inserted into the response for binding subsequent requests from the same user to that server. The validity of the cookie is based on the cookie expiration time, which is specified in the policy configuration.

For more information, see Duration-Based Session Stickiness in the Classic Load Balancers Guide.

Creating a Request

data CreateLBCookieStickinessPolicy

Contains the parameters for CreateLBCookieStickinessPolicy.

See: newCreateLBCookieStickinessPolicy smart constructor.

Constructors

CreateLBCookieStickinessPolicy'

Fields

cookieExpirationPeriod :: Maybe Integer

The time period, in seconds, after which the cookie should be considered stale. If you do not specify this parameter, the default value is 0, which indicates that the sticky session should last for the duration of the browser session.

loadBalancerName :: Text

The name of the load balancer.

policyName :: Text

The name of the policy being created. Policy names must consist of alphanumeric characters and dashes (-). This name must be unique within the set of policies for this load balancer.

newCreateLBCookieStickinessPolicy

Arguments

:: Text
$sel:loadBalancerName:CreateLBCookieStickinessPolicy'

-> Text
$sel:policyName:CreateLBCookieStickinessPolicy'

-> CreateLBCookieStickinessPolicy

Create a value of CreateLBCookieStickinessPolicy with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:cookieExpirationPeriod:CreateLBCookieStickinessPolicy', createLBCookieStickinessPolicy_cookieExpirationPeriod - The time period, in seconds, after which the cookie should be considered stale. If you do not specify this parameter, the default value is 0, which indicates that the sticky session should last for the duration of the browser session.

$sel:loadBalancerName:CreateLBCookieStickinessPolicy', createLBCookieStickinessPolicy_loadBalancerName - The name of the load balancer.

$sel:policyName:CreateLBCookieStickinessPolicy', createLBCookieStickinessPolicy_policyName - The name of the policy being created. Policy names must consist of alphanumeric characters and dashes (-). This name must be unique within the set of policies for this load balancer.

Request Lenses

createLBCookieStickinessPolicy_cookieExpirationPeriod :: Lens' CreateLBCookieStickinessPolicy (Maybe Integer)

The time period, in seconds, after which the cookie should be considered stale. If you do not specify this parameter, the default value is 0, which indicates that the sticky session should last for the duration of the browser session.

createLBCookieStickinessPolicy_loadBalancerName :: Lens' CreateLBCookieStickinessPolicy Text

The name of the load balancer.

createLBCookieStickinessPolicy_policyName :: Lens' CreateLBCookieStickinessPolicy Text

The name of the policy being created. Policy names must consist of alphanumeric characters and dashes (-). This name must be unique within the set of policies for this load balancer.

Destructuring the Response

data CreateLBCookieStickinessPolicyResponse

Contains the output for CreateLBCookieStickinessPolicy.

See: newCreateLBCookieStickinessPolicyResponse smart constructor.

Constructors

CreateLBCookieStickinessPolicyResponse'

Fields

httpStatus :: Int

The response's http status code.

newCreateLBCookieStickinessPolicyResponse

Arguments

:: Int
$sel:httpStatus:CreateLBCookieStickinessPolicyResponse'

-> CreateLBCookieStickinessPolicyResponse

Create a value of CreateLBCookieStickinessPolicyResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:httpStatus:CreateLBCookieStickinessPolicyResponse', createLBCookieStickinessPolicyResponse_httpStatus - The response's http status code.

Response Lenses

createLBCookieStickinessPolicyResponse_httpStatus :: Lens' CreateLBCookieStickinessPolicyResponse Int

The response's http status code.