Amazonka.ELB.DeregisterInstancesFromLoadBalancer

Description

Deregisters the specified instances from the specified load balancer. After the instance is deregistered, it no longer receives traffic from the load balancer.

You can use DescribeLoadBalancers to verify that the instance is deregistered from the load balancer.

For more information, see Register or De-Register EC2 Instances in the Classic Load Balancers Guide.

Creating a Request

data DeregisterInstancesFromLoadBalancer

Contains the parameters for DeregisterInstancesFromLoadBalancer.

See: newDeregisterInstancesFromLoadBalancer smart constructor.

Constructors

DeregisterInstancesFromLoadBalancer'

Fields

loadBalancerName :: Text

The name of the load balancer.

instances :: [Instance]

The IDs of the instances.

newDeregisterInstancesFromLoadBalancer

Arguments

:: Text
$sel:loadBalancerName:DeregisterInstancesFromLoadBalancer'

-> DeregisterInstancesFromLoadBalancer

Create a value of DeregisterInstancesFromLoadBalancer with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:loadBalancerName:DeregisterInstancesFromLoadBalancer', deregisterInstancesFromLoadBalancer_loadBalancerName - The name of the load balancer.

$sel:instances:DeregisterInstancesFromLoadBalancer', deregisterInstancesFromLoadBalancer_instances - The IDs of the instances.

Request Lenses

deregisterInstancesFromLoadBalancer_loadBalancerName :: Lens' DeregisterInstancesFromLoadBalancer Text

The name of the load balancer.

deregisterInstancesFromLoadBalancer_instances :: Lens' DeregisterInstancesFromLoadBalancer [Instance]

The IDs of the instances.

Destructuring the Response

data DeregisterInstancesFromLoadBalancerResponse

Contains the output of DeregisterInstancesFromLoadBalancer.

See: newDeregisterInstancesFromLoadBalancerResponse smart constructor.

Constructors

DeregisterInstancesFromLoadBalancerResponse'

Fields

instances :: Maybe [Instance]

The remaining instances registered with the load balancer.

httpStatus :: Int

The response's http status code.

newDeregisterInstancesFromLoadBalancerResponse

Arguments

:: Int
$sel:httpStatus:DeregisterInstancesFromLoadBalancerResponse'

-> DeregisterInstancesFromLoadBalancerResponse

Create a value of DeregisterInstancesFromLoadBalancerResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:instances:DeregisterInstancesFromLoadBalancer', deregisterInstancesFromLoadBalancerResponse_instances - The remaining instances registered with the load balancer.

$sel:httpStatus:DeregisterInstancesFromLoadBalancerResponse', deregisterInstancesFromLoadBalancerResponse_httpStatus - The response's http status code.

Response Lenses

deregisterInstancesFromLoadBalancerResponse_instances :: Lens' DeregisterInstancesFromLoadBalancerResponse (Maybe [Instance])

The remaining instances registered with the load balancer.

deregisterInstancesFromLoadBalancerResponse_httpStatus :: Lens' DeregisterInstancesFromLoadBalancerResponse Int

The response's http status code.