Amazonka.ELB.DescribeLoadBalancerPolicies

Description

Describes the specified policies.

If you specify a load balancer name, the action returns the descriptions of all policies created for the load balancer. If you specify a policy name associated with your load balancer, the action returns the description of that policy. If you don't specify a load balancer name, the action returns descriptions of the specified sample policies, or descriptions of all sample policies. The names of the sample policies have the ELBSample- prefix.

Creating a Request

data DescribeLoadBalancerPolicies

Contains the parameters for DescribeLoadBalancerPolicies.

See: newDescribeLoadBalancerPolicies smart constructor.

Constructors

DescribeLoadBalancerPolicies'

Fields

policyNames :: Maybe [Text]

The names of the policies.

loadBalancerName :: Maybe Text

The name of the load balancer.

newDescribeLoadBalancerPolicies :: DescribeLoadBalancerPolicies

Create a value of DescribeLoadBalancerPolicies with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:policyNames:DescribeLoadBalancerPolicies', describeLoadBalancerPolicies_policyNames - The names of the policies.

$sel:loadBalancerName:DescribeLoadBalancerPolicies', describeLoadBalancerPolicies_loadBalancerName - The name of the load balancer.

Request Lenses

describeLoadBalancerPolicies_policyNames :: Lens' DescribeLoadBalancerPolicies (Maybe [Text])

The names of the policies.

describeLoadBalancerPolicies_loadBalancerName :: Lens' DescribeLoadBalancerPolicies (Maybe Text)

The name of the load balancer.

Destructuring the Response

data DescribeLoadBalancerPoliciesResponse

Contains the output of DescribeLoadBalancerPolicies.

See: newDescribeLoadBalancerPoliciesResponse smart constructor.

Constructors

DescribeLoadBalancerPoliciesResponse'

Fields

policyDescriptions :: Maybe [PolicyDescription]

Information about the policies.

httpStatus :: Int

The response's http status code.

newDescribeLoadBalancerPoliciesResponse

Arguments

:: Int
$sel:httpStatus:DescribeLoadBalancerPoliciesResponse'

-> DescribeLoadBalancerPoliciesResponse

Create a value of DescribeLoadBalancerPoliciesResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:policyDescriptions:DescribeLoadBalancerPoliciesResponse', describeLoadBalancerPoliciesResponse_policyDescriptions - Information about the policies.

$sel:httpStatus:DescribeLoadBalancerPoliciesResponse', describeLoadBalancerPoliciesResponse_httpStatus - The response's http status code.

Response Lenses

describeLoadBalancerPoliciesResponse_policyDescriptions :: Lens' DescribeLoadBalancerPoliciesResponse (Maybe [PolicyDescription])

Information about the policies.

describeLoadBalancerPoliciesResponse_httpStatus :: Lens' DescribeLoadBalancerPoliciesResponse Int

The response's http status code.