Amazonka.ELB.Types.LoadBalancerAttributes

Description

Documentation

data LoadBalancerAttributes

The attributes for a load balancer.

See: newLoadBalancerAttributes smart constructor.

Constructors

LoadBalancerAttributes'

Fields

crossZoneLoadBalancing :: Maybe CrossZoneLoadBalancing

If enabled, the load balancer routes the request traffic evenly across all instances regardless of the Availability Zones.

For more information, see Configure Cross-Zone Load Balancing in the Classic Load Balancers Guide.

accessLog :: Maybe AccessLog

If enabled, the load balancer captures detailed information of all requests and delivers the information to the Amazon S3 bucket that you specify.

For more information, see Enable Access Logs in the Classic Load Balancers Guide.

additionalAttributes :: Maybe [AdditionalAttribute]

Any additional attributes.

connectionSettings :: Maybe ConnectionSettings

If enabled, the load balancer allows the connections to remain idle (no data is sent over the connection) for the specified duration.

By default, Elastic Load Balancing maintains a 60-second idle connection timeout for both front-end and back-end connections of your load balancer. For more information, see Configure Idle Connection Timeout in the Classic Load Balancers Guide.

connectionDraining :: Maybe ConnectionDraining

If enabled, the load balancer allows existing requests to complete before the load balancer shifts traffic away from a deregistered or unhealthy instance.

For more information, see Configure Connection Draining in the Classic Load Balancers Guide.

newLoadBalancerAttributes :: LoadBalancerAttributes

Create a value of LoadBalancerAttributes with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:crossZoneLoadBalancing:LoadBalancerAttributes', loadBalancerAttributes_crossZoneLoadBalancing - If enabled, the load balancer routes the request traffic evenly across all instances regardless of the Availability Zones.

For more information, see Configure Cross-Zone Load Balancing in the Classic Load Balancers Guide.

$sel:accessLog:LoadBalancerAttributes', loadBalancerAttributes_accessLog - If enabled, the load balancer captures detailed information of all requests and delivers the information to the Amazon S3 bucket that you specify.

For more information, see Enable Access Logs in the Classic Load Balancers Guide.

$sel:additionalAttributes:LoadBalancerAttributes', loadBalancerAttributes_additionalAttributes - Any additional attributes.

$sel:connectionSettings:LoadBalancerAttributes', loadBalancerAttributes_connectionSettings - If enabled, the load balancer allows the connections to remain idle (no data is sent over the connection) for the specified duration.

By default, Elastic Load Balancing maintains a 60-second idle connection timeout for both front-end and back-end connections of your load balancer. For more information, see Configure Idle Connection Timeout in the Classic Load Balancers Guide.

$sel:connectionDraining:LoadBalancerAttributes', loadBalancerAttributes_connectionDraining - If enabled, the load balancer allows existing requests to complete before the load balancer shifts traffic away from a deregistered or unhealthy instance.

For more information, see Configure Connection Draining in the Classic Load Balancers Guide.

loadBalancerAttributes_crossZoneLoadBalancing :: Lens' LoadBalancerAttributes (Maybe CrossZoneLoadBalancing)

If enabled, the load balancer routes the request traffic evenly across all instances regardless of the Availability Zones.

For more information, see Configure Cross-Zone Load Balancing in the Classic Load Balancers Guide.

loadBalancerAttributes_accessLog :: Lens' LoadBalancerAttributes (Maybe AccessLog)

If enabled, the load balancer captures detailed information of all requests and delivers the information to the Amazon S3 bucket that you specify.

For more information, see Enable Access Logs in the Classic Load Balancers Guide.

loadBalancerAttributes_additionalAttributes :: Lens' LoadBalancerAttributes (Maybe [AdditionalAttribute])

Any additional attributes.

loadBalancerAttributes_connectionSettings :: Lens' LoadBalancerAttributes (Maybe ConnectionSettings)

If enabled, the load balancer allows the connections to remain idle (no data is sent over the connection) for the specified duration.

By default, Elastic Load Balancing maintains a 60-second idle connection timeout for both front-end and back-end connections of your load balancer. For more information, see Configure Idle Connection Timeout in the Classic Load Balancers Guide.

loadBalancerAttributes_connectionDraining :: Lens' LoadBalancerAttributes (Maybe ConnectionDraining)

If enabled, the load balancer allows existing requests to complete before the load balancer shifts traffic away from a deregistered or unhealthy instance.

For more information, see Configure Connection Draining in the Classic Load Balancers Guide.