Amazonka.ELBV2.DescribeTargetGroupAttributes

Description

Describes the attributes for the specified target group.

For more information, see the following:

Target group attributes in the Application Load Balancers Guide

Target group attributes in the Network Load Balancers Guide

Target group attributes in the Gateway Load Balancers Guide

Creating a Request

data DescribeTargetGroupAttributes

See: newDescribeTargetGroupAttributes smart constructor.

Constructors

DescribeTargetGroupAttributes'

Fields

targetGroupArn :: Text

The Amazon Resource Name (ARN) of the target group.

newDescribeTargetGroupAttributes

Arguments

:: Text
$sel:targetGroupArn:DescribeTargetGroupAttributes'

-> DescribeTargetGroupAttributes

Create a value of DescribeTargetGroupAttributes with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:targetGroupArn:DescribeTargetGroupAttributes', describeTargetGroupAttributes_targetGroupArn - The Amazon Resource Name (ARN) of the target group.

Request Lenses

describeTargetGroupAttributes_targetGroupArn :: Lens' DescribeTargetGroupAttributes Text

The Amazon Resource Name (ARN) of the target group.

Destructuring the Response

data DescribeTargetGroupAttributesResponse

See: newDescribeTargetGroupAttributesResponse smart constructor.

Constructors

DescribeTargetGroupAttributesResponse'

Fields

attributes :: Maybe [TargetGroupAttribute]

Information about the target group attributes

httpStatus :: Int

The response's http status code.

newDescribeTargetGroupAttributesResponse

Arguments

:: Int
$sel:httpStatus:DescribeTargetGroupAttributesResponse'

-> DescribeTargetGroupAttributesResponse

Create a value of DescribeTargetGroupAttributesResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:attributes:DescribeTargetGroupAttributesResponse', describeTargetGroupAttributesResponse_attributes - Information about the target group attributes

$sel:httpStatus:DescribeTargetGroupAttributesResponse', describeTargetGroupAttributesResponse_httpStatus - The response's http status code.

Response Lenses

describeTargetGroupAttributesResponse_attributes :: Lens' DescribeTargetGroupAttributesResponse (Maybe [TargetGroupAttribute])

Information about the target group attributes

describeTargetGroupAttributesResponse_httpStatus :: Lens' DescribeTargetGroupAttributesResponse Int

The response's http status code.