Amazonka.ELBV2.SetIpAddressType

Description

Sets the type of IP addresses used by the subnets of the specified Application Load Balancer or Network Load Balancer.

Creating a Request

data SetIpAddressType

See: newSetIpAddressType smart constructor.

Constructors

SetIpAddressType'

Fields

loadBalancerArn :: Text

The Amazon Resource Name (ARN) of the load balancer.

ipAddressType :: IpAddressType

The IP address type. The possible values are ipv4 (for IPv4 addresses) and dualstack (for IPv4 and IPv6 addresses). Internal load balancers must use ipv4. You can’t specify dualstack for a load balancer with a UDP or TCP_UDP listener.

newSetIpAddressType

Arguments

:: Text
$sel:loadBalancerArn:SetIpAddressType'

-> IpAddressType
$sel:ipAddressType:SetIpAddressType'

-> SetIpAddressType

Create a value of SetIpAddressType with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:loadBalancerArn:SetIpAddressType', setIpAddressType_loadBalancerArn - The Amazon Resource Name (ARN) of the load balancer.

$sel:ipAddressType:SetIpAddressType', setIpAddressType_ipAddressType - The IP address type. The possible values are ipv4 (for IPv4 addresses) and dualstack (for IPv4 and IPv6 addresses). Internal load balancers must use ipv4. You can’t specify dualstack for a load balancer with a UDP or TCP_UDP listener.

Request Lenses

setIpAddressType_loadBalancerArn :: Lens' SetIpAddressType Text

The Amazon Resource Name (ARN) of the load balancer.

setIpAddressType_ipAddressType :: Lens' SetIpAddressType IpAddressType

The IP address type. The possible values are ipv4 (for IPv4 addresses) and dualstack (for IPv4 and IPv6 addresses). Internal load balancers must use ipv4. You can’t specify dualstack for a load balancer with a UDP or TCP_UDP listener.

Destructuring the Response

data SetIpAddressTypeResponse

See: newSetIpAddressTypeResponse smart constructor.

Constructors

SetIpAddressTypeResponse'

Fields

ipAddressType :: Maybe IpAddressType

The IP address type.

httpStatus :: Int

The response's http status code.

newSetIpAddressTypeResponse

Arguments

:: Int
$sel:httpStatus:SetIpAddressTypeResponse'

-> SetIpAddressTypeResponse

Create a value of SetIpAddressTypeResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:ipAddressType:SetIpAddressType', setIpAddressTypeResponse_ipAddressType - The IP address type.

$sel:httpStatus:SetIpAddressTypeResponse', setIpAddressTypeResponse_httpStatus - The response's http status code.

Response Lenses

setIpAddressTypeResponse_ipAddressType :: Lens' SetIpAddressTypeResponse (Maybe IpAddressType)

The IP address type.

setIpAddressTypeResponse_httpStatus :: Lens' SetIpAddressTypeResponse Int

The response's http status code.