Amazonka.ELBV2.SetSubnets

Description

Enables the Availability Zones for the specified public subnets for the specified Application Load Balancer or Network Load Balancer. The specified subnets replace the previously enabled subnets.

When you specify subnets for a Network Load Balancer, you must include all subnets that were enabled previously, with their existing configurations, plus any additional subnets.

Creating a Request

data SetSubnets

See: newSetSubnets smart constructor.

Constructors

SetSubnets'

Fields

subnetMappings :: Maybe [SubnetMapping]

The IDs of the public subnets. You can specify only one subnet per Availability Zone. You must specify either subnets or subnet mappings.

Application Load Balancers

You must specify subnets from at least two Availability Zones. You cannot specify Elastic IP addresses for your subnets.

Application Load Balancers on Outposts

You must specify one Outpost subnet.

Application Load Balancers on Local Zones

You can specify subnets from one or more Local Zones.

Network Load Balancers

You can specify subnets from one or more Availability Zones. You can specify one Elastic IP address per subnet if you need static IP addresses for your internet-facing load balancer. For internal load balancers, you can specify one private IP address per subnet from the IPv4 range of the subnet. For internet-facing load balancer, you can specify one IPv6 address per subnet.

subnets :: Maybe [Text]

The IDs of the public subnets. You can specify only one subnet per Availability Zone. You must specify either subnets or subnet mappings.

Application Load Balancers

You must specify subnets from at least two Availability Zones.

Application Load Balancers on Outposts

You must specify one Outpost subnet.

Application Load Balancers on Local Zones

You can specify subnets from one or more Local Zones.

Network Load Balancers

You can specify subnets from one or more Availability Zones.

ipAddressType :: Maybe IpAddressType

Network Load Balancers

The type of IP addresses used by the subnets for your load balancer. The possible values are ipv4 (for IPv4 addresses) and dualstack (for IPv4 and IPv6 addresses). You can’t specify dualstack for a load balancer with a UDP or TCP_UDP listener. Internal load balancers must use ipv4.

loadBalancerArn :: Text

The Amazon Resource Name (ARN) of the load balancer.

newSetSubnets

Arguments

:: Text
$sel:loadBalancerArn:SetSubnets'

-> SetSubnets

Create a value of SetSubnets with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:subnetMappings:SetSubnets', setSubnets_subnetMappings - The IDs of the public subnets. You can specify only one subnet per Availability Zone. You must specify either subnets or subnet mappings.

Application Load Balancers

You must specify subnets from at least two Availability Zones. You cannot specify Elastic IP addresses for your subnets.

Application Load Balancers on Outposts

You must specify one Outpost subnet.

Application Load Balancers on Local Zones

You can specify subnets from one or more Local Zones.

Network Load Balancers

You can specify subnets from one or more Availability Zones. You can specify one Elastic IP address per subnet if you need static IP addresses for your internet-facing load balancer. For internal load balancers, you can specify one private IP address per subnet from the IPv4 range of the subnet. For internet-facing load balancer, you can specify one IPv6 address per subnet.

$sel:subnets:SetSubnets', setSubnets_subnets - The IDs of the public subnets. You can specify only one subnet per Availability Zone. You must specify either subnets or subnet mappings.

Application Load Balancers

You must specify subnets from at least two Availability Zones.

Application Load Balancers on Outposts

You must specify one Outpost subnet.

Application Load Balancers on Local Zones

You can specify subnets from one or more Local Zones.

Network Load Balancers

You can specify subnets from one or more Availability Zones.

$sel:ipAddressType:SetSubnets', setSubnets_ipAddressType - [Network Load Balancers] The type of IP addresses used by the subnets for your load balancer. The possible values are ipv4 (for IPv4 addresses) and dualstack (for IPv4 and IPv6 addresses). You can’t specify dualstack for a load balancer with a UDP or TCP_UDP listener. Internal load balancers must use ipv4.

$sel:loadBalancerArn:SetSubnets', setSubnets_loadBalancerArn - The Amazon Resource Name (ARN) of the load balancer.

Request Lenses

setSubnets_subnetMappings :: Lens' SetSubnets (Maybe [SubnetMapping])

The IDs of the public subnets. You can specify only one subnet per Availability Zone. You must specify either subnets or subnet mappings.

Application Load Balancers

You must specify subnets from at least two Availability Zones. You cannot specify Elastic IP addresses for your subnets.

Application Load Balancers on Outposts

You must specify one Outpost subnet.

Application Load Balancers on Local Zones

You can specify subnets from one or more Local Zones.

Network Load Balancers

You can specify subnets from one or more Availability Zones. You can specify one Elastic IP address per subnet if you need static IP addresses for your internet-facing load balancer. For internal load balancers, you can specify one private IP address per subnet from the IPv4 range of the subnet. For internet-facing load balancer, you can specify one IPv6 address per subnet.

setSubnets_subnets :: Lens' SetSubnets (Maybe [Text])

The IDs of the public subnets. You can specify only one subnet per Availability Zone. You must specify either subnets or subnet mappings.

Application Load Balancers

You must specify subnets from at least two Availability Zones.

Application Load Balancers on Outposts

You must specify one Outpost subnet.

Application Load Balancers on Local Zones

You can specify subnets from one or more Local Zones.

Network Load Balancers

You can specify subnets from one or more Availability Zones.

setSubnets_ipAddressType :: Lens' SetSubnets (Maybe IpAddressType)

Network Load Balancers

The type of IP addresses used by the subnets for your load balancer. The possible values are ipv4 (for IPv4 addresses) and dualstack (for IPv4 and IPv6 addresses). You can’t specify dualstack for a load balancer with a UDP or TCP_UDP listener. Internal load balancers must use ipv4.

setSubnets_loadBalancerArn :: Lens' SetSubnets Text

The Amazon Resource Name (ARN) of the load balancer.

Destructuring the Response

data SetSubnetsResponse

See: newSetSubnetsResponse smart constructor.

Constructors

SetSubnetsResponse'

Fields

availabilityZones :: Maybe [AvailabilityZone]

Information about the subnets.

ipAddressType :: Maybe IpAddressType

Network Load Balancers

The IP address type.

httpStatus :: Int

The response's http status code.

newSetSubnetsResponse

Arguments

:: Int
$sel:httpStatus:SetSubnetsResponse'

-> SetSubnetsResponse

Create a value of SetSubnetsResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:availabilityZones:SetSubnetsResponse', setSubnetsResponse_availabilityZones - Information about the subnets.

$sel:ipAddressType:SetSubnets', setSubnetsResponse_ipAddressType - [Network Load Balancers] The IP address type.

$sel:httpStatus:SetSubnetsResponse', setSubnetsResponse_httpStatus - The response's http status code.

Response Lenses

setSubnetsResponse_availabilityZones :: Lens' SetSubnetsResponse (Maybe [AvailabilityZone])

Information about the subnets.

setSubnetsResponse_ipAddressType :: Lens' SetSubnetsResponse (Maybe IpAddressType)

Network Load Balancers

The IP address type.

setSubnetsResponse_httpStatus :: Lens' SetSubnetsResponse Int

The response's http status code.