Amazonka.ELBV2.Types.Rule

Description

Documentation

data Rule

Information about a rule.

See: newRule smart constructor.

Constructors

Rule'

Fields

priority :: Maybe Text

The priority.

actions :: Maybe [Action]

The actions. Each rule must include exactly one of the following types of actions: forward, redirect, or fixed-response, and it must be the last action to be performed.

conditions :: Maybe [RuleCondition]

The conditions. Each rule can include zero or one of the following conditions: http-request-method, host-header, path-pattern, and source-ip, and zero or more of the following conditions: http-header and query-string.

ruleArn :: Maybe Text

The Amazon Resource Name (ARN) of the rule.

isDefault :: Maybe Bool

Indicates whether this is the default rule.

newRule :: Rule

Create a value of Rule with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:priority:Rule', rule_priority - The priority.

$sel:actions:Rule', rule_actions - The actions. Each rule must include exactly one of the following types of actions: forward, redirect, or fixed-response, and it must be the last action to be performed.

$sel:conditions:Rule', rule_conditions - The conditions. Each rule can include zero or one of the following conditions: http-request-method, host-header, path-pattern, and source-ip, and zero or more of the following conditions: http-header and query-string.

$sel:ruleArn:Rule', rule_ruleArn - The Amazon Resource Name (ARN) of the rule.

$sel:isDefault:Rule', rule_isDefault - Indicates whether this is the default rule.

rule_priority :: Lens' Rule (Maybe Text)

The priority.

rule_actions :: Lens' Rule (Maybe [Action])

The actions. Each rule must include exactly one of the following types of actions: forward, redirect, or fixed-response, and it must be the last action to be performed.

rule_conditions :: Lens' Rule (Maybe [RuleCondition])

The conditions. Each rule can include zero or one of the following conditions: http-request-method, host-header, path-pattern, and source-ip, and zero or more of the following conditions: http-header and query-string.

rule_ruleArn :: Lens' Rule (Maybe Text)

The Amazon Resource Name (ARN) of the rule.

rule_isDefault :: Lens' Rule (Maybe Bool)

Indicates whether this is the default rule.