Amazonka.EMR.CreateStudio

Description

Creates a new Amazon EMR Studio.

Creating a Request

data CreateStudio

See: newCreateStudio smart constructor.

Constructors

CreateStudio'

Fields

idpAuthUrl :: Maybe Text

The authentication endpoint of your identity provider (IdP). Specify this value when you use IAM authentication and want to let federated users log in to a Studio with the Studio URL and credentials from your IdP. Amazon EMR Studio redirects users to this endpoint to enter credentials.

idpRelayStateParameterName :: Maybe Text

The name that your identity provider (IdP) uses for its RelayState parameter. For example, RelayState or TargetSource. Specify this value when you use IAM authentication and want to let federated users log in to a Studio using the Studio URL. The RelayState parameter differs by IdP.

userRole :: Maybe Text

The IAM user role that users and groups assume when logged in to an Amazon EMR Studio. Only specify a UserRole when you use Amazon Web Services SSO authentication. The permissions attached to the UserRole can be scoped down for each user or group using session policies.

description :: Maybe Text

A detailed description of the Amazon EMR Studio.

tags :: Maybe [Tag]

A list of tags to associate with the Amazon EMR Studio. Tags are user-defined key-value pairs that consist of a required key string with a maximum of 128 characters, and an optional value string with a maximum of 256 characters.

name :: Text

A descriptive name for the Amazon EMR Studio.

authMode :: AuthMode

Specifies whether the Studio authenticates users using IAM or Amazon Web Services SSO.

vpcId :: Text

The ID of the Amazon Virtual Private Cloud (Amazon VPC) to associate with the Studio.

subnetIds :: [Text]

A list of subnet IDs to associate with the Amazon EMR Studio. A Studio can have a maximum of 5 subnets. The subnets must belong to the VPC specified by VpcId. Studio users can create a Workspace in any of the specified subnets.

serviceRole :: Text

The IAM role that the Amazon EMR Studio assumes. The service role provides a way for Amazon EMR Studio to interoperate with other Amazon Web Services services.

workspaceSecurityGroupId :: Text

The ID of the Amazon EMR Studio Workspace security group. The Workspace security group allows outbound network traffic to resources in the Engine security group, and it must be in the same VPC specified by VpcId.

engineSecurityGroupId :: Text

The ID of the Amazon EMR Studio Engine security group. The Engine security group allows inbound network traffic from the Workspace security group, and it must be in the same VPC specified by VpcId.

defaultS3Location :: Text

The Amazon S3 location to back up Amazon EMR Studio Workspaces and notebook files.

newCreateStudio

Arguments

:: Text
$sel:name:CreateStudio'

-> AuthMode
$sel:authMode:CreateStudio'

-> Text
$sel:vpcId:CreateStudio'

-> Text
$sel:serviceRole:CreateStudio'

-> Text
$sel:workspaceSecurityGroupId:CreateStudio'

-> Text
$sel:engineSecurityGroupId:CreateStudio'

-> Text
$sel:defaultS3Location:CreateStudio'

-> CreateStudio

Create a value of CreateStudio with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:idpAuthUrl:CreateStudio', createStudio_idpAuthUrl - The authentication endpoint of your identity provider (IdP). Specify this value when you use IAM authentication and want to let federated users log in to a Studio with the Studio URL and credentials from your IdP. Amazon EMR Studio redirects users to this endpoint to enter credentials.

$sel:idpRelayStateParameterName:CreateStudio', createStudio_idpRelayStateParameterName - The name that your identity provider (IdP) uses for its RelayState parameter. For example, RelayState or TargetSource. Specify this value when you use IAM authentication and want to let federated users log in to a Studio using the Studio URL. The RelayState parameter differs by IdP.

$sel:userRole:CreateStudio', createStudio_userRole - The IAM user role that users and groups assume when logged in to an Amazon EMR Studio. Only specify a UserRole when you use Amazon Web Services SSO authentication. The permissions attached to the UserRole can be scoped down for each user or group using session policies.

$sel:description:CreateStudio', createStudio_description - A detailed description of the Amazon EMR Studio.

$sel:tags:CreateStudio', createStudio_tags - A list of tags to associate with the Amazon EMR Studio. Tags are user-defined key-value pairs that consist of a required key string with a maximum of 128 characters, and an optional value string with a maximum of 256 characters.

$sel:name:CreateStudio', createStudio_name - A descriptive name for the Amazon EMR Studio.

$sel:authMode:CreateStudio', createStudio_authMode - Specifies whether the Studio authenticates users using IAM or Amazon Web Services SSO.

$sel:vpcId:CreateStudio', createStudio_vpcId - The ID of the Amazon Virtual Private Cloud (Amazon VPC) to associate with the Studio.

$sel:subnetIds:CreateStudio', createStudio_subnetIds - A list of subnet IDs to associate with the Amazon EMR Studio. A Studio can have a maximum of 5 subnets. The subnets must belong to the VPC specified by VpcId. Studio users can create a Workspace in any of the specified subnets.

$sel:serviceRole:CreateStudio', createStudio_serviceRole - The IAM role that the Amazon EMR Studio assumes. The service role provides a way for Amazon EMR Studio to interoperate with other Amazon Web Services services.

$sel:workspaceSecurityGroupId:CreateStudio', createStudio_workspaceSecurityGroupId - The ID of the Amazon EMR Studio Workspace security group. The Workspace security group allows outbound network traffic to resources in the Engine security group, and it must be in the same VPC specified by VpcId.

$sel:engineSecurityGroupId:CreateStudio', createStudio_engineSecurityGroupId - The ID of the Amazon EMR Studio Engine security group. The Engine security group allows inbound network traffic from the Workspace security group, and it must be in the same VPC specified by VpcId.

$sel:defaultS3Location:CreateStudio', createStudio_defaultS3Location - The Amazon S3 location to back up Amazon EMR Studio Workspaces and notebook files.

Request Lenses

createStudio_idpAuthUrl :: Lens' CreateStudio (Maybe Text)

The authentication endpoint of your identity provider (IdP). Specify this value when you use IAM authentication and want to let federated users log in to a Studio with the Studio URL and credentials from your IdP. Amazon EMR Studio redirects users to this endpoint to enter credentials.

createStudio_idpRelayStateParameterName :: Lens' CreateStudio (Maybe Text)

The name that your identity provider (IdP) uses for its RelayState parameter. For example, RelayState or TargetSource. Specify this value when you use IAM authentication and want to let federated users log in to a Studio using the Studio URL. The RelayState parameter differs by IdP.

createStudio_userRole :: Lens' CreateStudio (Maybe Text)

The IAM user role that users and groups assume when logged in to an Amazon EMR Studio. Only specify a UserRole when you use Amazon Web Services SSO authentication. The permissions attached to the UserRole can be scoped down for each user or group using session policies.

createStudio_description :: Lens' CreateStudio (Maybe Text)

A detailed description of the Amazon EMR Studio.

createStudio_tags :: Lens' CreateStudio (Maybe [Tag])

A list of tags to associate with the Amazon EMR Studio. Tags are user-defined key-value pairs that consist of a required key string with a maximum of 128 characters, and an optional value string with a maximum of 256 characters.

createStudio_name :: Lens' CreateStudio Text

A descriptive name for the Amazon EMR Studio.

createStudio_authMode :: Lens' CreateStudio AuthMode

Specifies whether the Studio authenticates users using IAM or Amazon Web Services SSO.

createStudio_vpcId :: Lens' CreateStudio Text

The ID of the Amazon Virtual Private Cloud (Amazon VPC) to associate with the Studio.

createStudio_subnetIds :: Lens' CreateStudio [Text]

A list of subnet IDs to associate with the Amazon EMR Studio. A Studio can have a maximum of 5 subnets. The subnets must belong to the VPC specified by VpcId. Studio users can create a Workspace in any of the specified subnets.

createStudio_serviceRole :: Lens' CreateStudio Text

The IAM role that the Amazon EMR Studio assumes. The service role provides a way for Amazon EMR Studio to interoperate with other Amazon Web Services services.

createStudio_workspaceSecurityGroupId :: Lens' CreateStudio Text

The ID of the Amazon EMR Studio Workspace security group. The Workspace security group allows outbound network traffic to resources in the Engine security group, and it must be in the same VPC specified by VpcId.

createStudio_engineSecurityGroupId :: Lens' CreateStudio Text

The ID of the Amazon EMR Studio Engine security group. The Engine security group allows inbound network traffic from the Workspace security group, and it must be in the same VPC specified by VpcId.

createStudio_defaultS3Location :: Lens' CreateStudio Text

The Amazon S3 location to back up Amazon EMR Studio Workspaces and notebook files.

Destructuring the Response

data CreateStudioResponse

See: newCreateStudioResponse smart constructor.

Constructors

CreateStudioResponse'

Fields

studioId :: Maybe Text

The ID of the Amazon EMR Studio.

url :: Maybe Text

The unique Studio access URL.

httpStatus :: Int

The response's http status code.

newCreateStudioResponse

Arguments

:: Int
$sel:httpStatus:CreateStudioResponse'

-> CreateStudioResponse

Create a value of CreateStudioResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:studioId:CreateStudioResponse', createStudioResponse_studioId - The ID of the Amazon EMR Studio.

$sel:url:CreateStudioResponse', createStudioResponse_url - The unique Studio access URL.

$sel:httpStatus:CreateStudioResponse', createStudioResponse_httpStatus - The response's http status code.

Response Lenses

createStudioResponse_studioId :: Lens' CreateStudioResponse (Maybe Text)

The ID of the Amazon EMR Studio.

createStudioResponse_url :: Lens' CreateStudioResponse (Maybe Text)

The unique Studio access URL.

createStudioResponse_httpStatus :: Lens' CreateStudioResponse Int

The response's http status code.