Amazonka.EMR.ListSteps

Description

Provides a list of steps for the cluster in reverse order unless you specify stepIds with the request or filter by StepStates. You can specify a maximum of 10 stepIDs. The CLI automatically paginates results to return a list greater than 50 steps. To return more than 50 steps using the CLI, specify a Marker, which is a pagination token that indicates the next set of steps to retrieve.

This operation returns paginated results.

Creating a Request

data ListSteps

This input determines which steps to list.

See: newListSteps smart constructor.

Constructors

ListSteps'

Fields

stepIds :: Maybe [Text]

The filter to limit the step list based on the identifier of the steps. You can specify a maximum of ten Step IDs. The character constraint applies to the overall length of the array.

stepStates :: Maybe [StepState]

The filter to limit the step list based on certain states.

marker :: Maybe Text

The maximum number of steps that a single ListSteps action returns is 50. To return a longer list of steps, use multiple ListSteps actions along with the Marker parameter, which is a pagination token that indicates the next set of results to retrieve.

clusterId :: Text

The identifier of the cluster for which to list the steps.

newListSteps

Arguments

:: Text
$sel:clusterId:ListSteps'

-> ListSteps

Create a value of ListSteps with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:stepIds:ListSteps', listSteps_stepIds - The filter to limit the step list based on the identifier of the steps. You can specify a maximum of ten Step IDs. The character constraint applies to the overall length of the array.

$sel:stepStates:ListSteps', listSteps_stepStates - The filter to limit the step list based on certain states.

$sel:marker:ListSteps', listSteps_marker - The maximum number of steps that a single ListSteps action returns is 50. To return a longer list of steps, use multiple ListSteps actions along with the Marker parameter, which is a pagination token that indicates the next set of results to retrieve.

$sel:clusterId:ListSteps', listSteps_clusterId - The identifier of the cluster for which to list the steps.

Request Lenses

listSteps_stepIds :: Lens' ListSteps (Maybe [Text])

The filter to limit the step list based on the identifier of the steps. You can specify a maximum of ten Step IDs. The character constraint applies to the overall length of the array.

listSteps_stepStates :: Lens' ListSteps (Maybe [StepState])

The filter to limit the step list based on certain states.

listSteps_marker :: Lens' ListSteps (Maybe Text)

The maximum number of steps that a single ListSteps action returns is 50. To return a longer list of steps, use multiple ListSteps actions along with the Marker parameter, which is a pagination token that indicates the next set of results to retrieve.

listSteps_clusterId :: Lens' ListSteps Text

The identifier of the cluster for which to list the steps.

Destructuring the Response

data ListStepsResponse

This output contains the list of steps returned in reverse order. This means that the last step is the first element in the list.

See: newListStepsResponse smart constructor.

Constructors

ListStepsResponse'

Fields

steps :: Maybe [StepSummary]

The filtered list of steps for the cluster.

marker :: Maybe Text

The maximum number of steps that a single ListSteps action returns is 50. To return a longer list of steps, use multiple ListSteps actions along with the Marker parameter, which is a pagination token that indicates the next set of results to retrieve.

httpStatus :: Int

The response's http status code.

newListStepsResponse

Arguments

:: Int
$sel:httpStatus:ListStepsResponse'

-> ListStepsResponse

Create a value of ListStepsResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:steps:ListStepsResponse', listStepsResponse_steps - The filtered list of steps for the cluster.

$sel:marker:ListSteps', listStepsResponse_marker - The maximum number of steps that a single ListSteps action returns is 50. To return a longer list of steps, use multiple ListSteps actions along with the Marker parameter, which is a pagination token that indicates the next set of results to retrieve.

$sel:httpStatus:ListStepsResponse', listStepsResponse_httpStatus - The response's http status code.

Response Lenses

listStepsResponse_steps :: Lens' ListStepsResponse (Maybe [StepSummary])

The filtered list of steps for the cluster.

listStepsResponse_marker :: Lens' ListStepsResponse (Maybe Text)

The maximum number of steps that a single ListSteps action returns is 50. To return a longer list of steps, use multiple ListSteps actions along with the Marker parameter, which is a pagination token that indicates the next set of results to retrieve.

listStepsResponse_httpStatus :: Lens' ListStepsResponse Int

The response's http status code.