Amazonka.EMR.ListStudioSessionMappings

Description

Returns a list of all user or group session mappings for the Amazon EMR Studio specified by StudioId.

This operation returns paginated results.

Creating a Request

data ListStudioSessionMappings

See: newListStudioSessionMappings smart constructor.

Constructors

ListStudioSessionMappings'

Fields

studioId :: Maybe Text

The ID of the Amazon EMR Studio.

identityType :: Maybe IdentityType

Specifies whether to return session mappings for users or groups. If not specified, the results include session mapping details for both users and groups.

marker :: Maybe Text

The pagination token that indicates the set of results to retrieve.

newListStudioSessionMappings :: ListStudioSessionMappings

Create a value of ListStudioSessionMappings with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:studioId:ListStudioSessionMappings', listStudioSessionMappings_studioId - The ID of the Amazon EMR Studio.

$sel:identityType:ListStudioSessionMappings', listStudioSessionMappings_identityType - Specifies whether to return session mappings for users or groups. If not specified, the results include session mapping details for both users and groups.

$sel:marker:ListStudioSessionMappings', listStudioSessionMappings_marker - The pagination token that indicates the set of results to retrieve.

Request Lenses

listStudioSessionMappings_studioId :: Lens' ListStudioSessionMappings (Maybe Text)

The ID of the Amazon EMR Studio.

listStudioSessionMappings_identityType :: Lens' ListStudioSessionMappings (Maybe IdentityType)

Specifies whether to return session mappings for users or groups. If not specified, the results include session mapping details for both users and groups.

listStudioSessionMappings_marker :: Lens' ListStudioSessionMappings (Maybe Text)

The pagination token that indicates the set of results to retrieve.

Destructuring the Response

data ListStudioSessionMappingsResponse

See: newListStudioSessionMappingsResponse smart constructor.

Constructors

ListStudioSessionMappingsResponse'

Fields

sessionMappings :: Maybe [SessionMappingSummary]

A list of session mapping summary objects. Each object includes session mapping details such as creation time, identity type (user or group), and Amazon EMR Studio ID.

marker :: Maybe Text

The pagination token that indicates the next set of results to retrieve.

httpStatus :: Int

The response's http status code.

newListStudioSessionMappingsResponse

Arguments

:: Int
$sel:httpStatus:ListStudioSessionMappingsResponse'

-> ListStudioSessionMappingsResponse

Create a value of ListStudioSessionMappingsResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:sessionMappings:ListStudioSessionMappingsResponse', listStudioSessionMappingsResponse_sessionMappings - A list of session mapping summary objects. Each object includes session mapping details such as creation time, identity type (user or group), and Amazon EMR Studio ID.

$sel:marker:ListStudioSessionMappings', listStudioSessionMappingsResponse_marker - The pagination token that indicates the next set of results to retrieve.

$sel:httpStatus:ListStudioSessionMappingsResponse', listStudioSessionMappingsResponse_httpStatus - The response's http status code.

Response Lenses

listStudioSessionMappingsResponse_sessionMappings :: Lens' ListStudioSessionMappingsResponse (Maybe [SessionMappingSummary])

A list of session mapping summary objects. Each object includes session mapping details such as creation time, identity type (user or group), and Amazon EMR Studio ID.

listStudioSessionMappingsResponse_marker :: Lens' ListStudioSessionMappingsResponse (Maybe Text)

The pagination token that indicates the next set of results to retrieve.

listStudioSessionMappingsResponse_httpStatus :: Lens' ListStudioSessionMappingsResponse Int

The response's http status code.