Amazonka.EMR.StartNotebookExecution

Description

Starts a notebook execution.

Creating a Request

data StartNotebookExecution

See: newStartNotebookExecution smart constructor.

Constructors

StartNotebookExecution'

Fields

notebookInstanceSecurityGroupId :: Maybe Text

The unique identifier of the Amazon EC2 security group to associate with the EMR Notebook for this notebook execution.

notebookExecutionName :: Maybe Text

An optional name for the notebook execution.

notebookParams :: Maybe Text

Input parameters in JSON format passed to the EMR Notebook at runtime for execution.

tags :: Maybe [Tag]

A list of tags associated with a notebook execution. Tags are user-defined key-value pairs that consist of a required key string with a maximum of 128 characters and an optional value string with a maximum of 256 characters.

editorId :: Text

The unique identifier of the EMR Notebook to use for notebook execution.

relativePath :: Text

The path and file name of the notebook file for this execution, relative to the path specified for the EMR Notebook. For example, if you specify a path of s3://MyBucket/MyNotebooks when you create an EMR Notebook for a notebook with an ID of e-ABCDEFGHIJK1234567890ABCD (the EditorID of this request), and you specify a RelativePath of my_notebook_executions/notebook_execution.ipynb, the location of the file for the notebook execution is s3://MyBucket/MyNotebooks/e-ABCDEFGHIJK1234567890ABCD/my_notebook_executions/notebook_execution.ipynb.

executionEngine :: ExecutionEngineConfig

Specifies the execution engine (cluster) that runs the notebook execution.

serviceRole :: Text

The name or ARN of the IAM role that is used as the service role for Amazon EMR (the EMR role) for the notebook execution.

newStartNotebookExecution

Arguments

:: Text
$sel:editorId:StartNotebookExecution'

-> Text
$sel:relativePath:StartNotebookExecution'

-> ExecutionEngineConfig
$sel:executionEngine:StartNotebookExecution'

-> Text
$sel:serviceRole:StartNotebookExecution'

-> StartNotebookExecution

Create a value of StartNotebookExecution with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:notebookInstanceSecurityGroupId:StartNotebookExecution', startNotebookExecution_notebookInstanceSecurityGroupId - The unique identifier of the Amazon EC2 security group to associate with the EMR Notebook for this notebook execution.

$sel:notebookExecutionName:StartNotebookExecution', startNotebookExecution_notebookExecutionName - An optional name for the notebook execution.

$sel:notebookParams:StartNotebookExecution', startNotebookExecution_notebookParams - Input parameters in JSON format passed to the EMR Notebook at runtime for execution.

$sel:tags:StartNotebookExecution', startNotebookExecution_tags - A list of tags associated with a notebook execution. Tags are user-defined key-value pairs that consist of a required key string with a maximum of 128 characters and an optional value string with a maximum of 256 characters.

$sel:editorId:StartNotebookExecution', startNotebookExecution_editorId - The unique identifier of the EMR Notebook to use for notebook execution.

$sel:relativePath:StartNotebookExecution', startNotebookExecution_relativePath - The path and file name of the notebook file for this execution, relative to the path specified for the EMR Notebook. For example, if you specify a path of s3://MyBucket/MyNotebooks when you create an EMR Notebook for a notebook with an ID of e-ABCDEFGHIJK1234567890ABCD (the EditorID of this request), and you specify a RelativePath of my_notebook_executions/notebook_execution.ipynb, the location of the file for the notebook execution is s3://MyBucket/MyNotebooks/e-ABCDEFGHIJK1234567890ABCD/my_notebook_executions/notebook_execution.ipynb.

$sel:executionEngine:StartNotebookExecution', startNotebookExecution_executionEngine - Specifies the execution engine (cluster) that runs the notebook execution.

$sel:serviceRole:StartNotebookExecution', startNotebookExecution_serviceRole - The name or ARN of the IAM role that is used as the service role for Amazon EMR (the EMR role) for the notebook execution.

Request Lenses

startNotebookExecution_notebookInstanceSecurityGroupId :: Lens' StartNotebookExecution (Maybe Text)

The unique identifier of the Amazon EC2 security group to associate with the EMR Notebook for this notebook execution.

startNotebookExecution_notebookExecutionName :: Lens' StartNotebookExecution (Maybe Text)

An optional name for the notebook execution.

startNotebookExecution_notebookParams :: Lens' StartNotebookExecution (Maybe Text)

Input parameters in JSON format passed to the EMR Notebook at runtime for execution.

startNotebookExecution_tags :: Lens' StartNotebookExecution (Maybe [Tag])

A list of tags associated with a notebook execution. Tags are user-defined key-value pairs that consist of a required key string with a maximum of 128 characters and an optional value string with a maximum of 256 characters.

startNotebookExecution_editorId :: Lens' StartNotebookExecution Text

The unique identifier of the EMR Notebook to use for notebook execution.

startNotebookExecution_relativePath :: Lens' StartNotebookExecution Text

The path and file name of the notebook file for this execution, relative to the path specified for the EMR Notebook. For example, if you specify a path of s3://MyBucket/MyNotebooks when you create an EMR Notebook for a notebook with an ID of e-ABCDEFGHIJK1234567890ABCD (the EditorID of this request), and you specify a RelativePath of my_notebook_executions/notebook_execution.ipynb, the location of the file for the notebook execution is s3://MyBucket/MyNotebooks/e-ABCDEFGHIJK1234567890ABCD/my_notebook_executions/notebook_execution.ipynb.

startNotebookExecution_executionEngine :: Lens' StartNotebookExecution ExecutionEngineConfig

Specifies the execution engine (cluster) that runs the notebook execution.

startNotebookExecution_serviceRole :: Lens' StartNotebookExecution Text

The name or ARN of the IAM role that is used as the service role for Amazon EMR (the EMR role) for the notebook execution.

Destructuring the Response

data StartNotebookExecutionResponse

See: newStartNotebookExecutionResponse smart constructor.

Constructors

StartNotebookExecutionResponse'

Fields

notebookExecutionId :: Maybe Text

The unique identifier of the notebook execution.

httpStatus :: Int

The response's http status code.

newStartNotebookExecutionResponse

Arguments

:: Int
$sel:httpStatus:StartNotebookExecutionResponse'

-> StartNotebookExecutionResponse

Create a value of StartNotebookExecutionResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:notebookExecutionId:StartNotebookExecutionResponse', startNotebookExecutionResponse_notebookExecutionId - The unique identifier of the notebook execution.

$sel:httpStatus:StartNotebookExecutionResponse', startNotebookExecutionResponse_httpStatus - The response's http status code.

Response Lenses

startNotebookExecutionResponse_notebookExecutionId :: Lens' StartNotebookExecutionResponse (Maybe Text)

The unique identifier of the notebook execution.

startNotebookExecutionResponse_httpStatus :: Lens' StartNotebookExecutionResponse Int

The response's http status code.