Amazonka.EMR.Types.AutoScalingPolicy

Description

Documentation

data AutoScalingPolicy

An automatic scaling policy for a core instance group or task instance group in an Amazon EMR cluster. An automatic scaling policy defines how an instance group dynamically adds and terminates EC2 instances in response to the value of a CloudWatch metric. See PutAutoScalingPolicy.

See: newAutoScalingPolicy smart constructor.

Constructors

AutoScalingPolicy'

Fields

constraints :: ScalingConstraints

The upper and lower EC2 instance limits for an automatic scaling policy. Automatic scaling activity will not cause an instance group to grow above or below these limits.

rules :: [ScalingRule]

The scale-in and scale-out rules that comprise the automatic scaling policy.

newAutoScalingPolicy

Arguments

:: ScalingConstraints
$sel:constraints:AutoScalingPolicy'

-> AutoScalingPolicy

Create a value of AutoScalingPolicy with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:constraints:AutoScalingPolicy', autoScalingPolicy_constraints - The upper and lower EC2 instance limits for an automatic scaling policy. Automatic scaling activity will not cause an instance group to grow above or below these limits.

$sel:rules:AutoScalingPolicy', autoScalingPolicy_rules - The scale-in and scale-out rules that comprise the automatic scaling policy.

autoScalingPolicy_constraints :: Lens' AutoScalingPolicy ScalingConstraints

The upper and lower EC2 instance limits for an automatic scaling policy. Automatic scaling activity will not cause an instance group to grow above or below these limits.

autoScalingPolicy_rules :: Lens' AutoScalingPolicy [ScalingRule]

The scale-in and scale-out rules that comprise the automatic scaling policy.