Amazonka.EMR.Types.AutoScalingPolicyDescription

Description

Documentation

data AutoScalingPolicyDescription

An automatic scaling policy for a core instance group or task instance group in an Amazon EMR cluster. The automatic scaling policy defines how an instance group dynamically adds and terminates EC2 instances in response to the value of a CloudWatch metric. See PutAutoScalingPolicy.

See: newAutoScalingPolicyDescription smart constructor.

Constructors

AutoScalingPolicyDescription'

Fields

status :: Maybe AutoScalingPolicyStatus

The status of an automatic scaling policy.

rules :: Maybe [ScalingRule]

The scale-in and scale-out rules that comprise the automatic scaling policy.

constraints :: Maybe ScalingConstraints

The upper and lower EC2 instance limits for an automatic scaling policy. Automatic scaling activity will not cause an instance group to grow above or below these limits.

newAutoScalingPolicyDescription :: AutoScalingPolicyDescription

Create a value of AutoScalingPolicyDescription with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:status:AutoScalingPolicyDescription', autoScalingPolicyDescription_status - The status of an automatic scaling policy.

$sel:rules:AutoScalingPolicyDescription', autoScalingPolicyDescription_rules - The scale-in and scale-out rules that comprise the automatic scaling policy.

$sel:constraints:AutoScalingPolicyDescription', autoScalingPolicyDescription_constraints - The upper and lower EC2 instance limits for an automatic scaling policy. Automatic scaling activity will not cause an instance group to grow above or below these limits.

autoScalingPolicyDescription_status :: Lens' AutoScalingPolicyDescription (Maybe AutoScalingPolicyStatus)

The status of an automatic scaling policy.

autoScalingPolicyDescription_rules :: Lens' AutoScalingPolicyDescription (Maybe [ScalingRule])

The scale-in and scale-out rules that comprise the automatic scaling policy.

autoScalingPolicyDescription_constraints :: Lens' AutoScalingPolicyDescription (Maybe ScalingConstraints)

The upper and lower EC2 instance limits for an automatic scaling policy. Automatic scaling activity will not cause an instance group to grow above or below these limits.