Amazonka.EMR.Types.Cluster

Description

Documentation

data Cluster

The detailed description of the cluster.

See: newCluster smart constructor.

Constructors

Cluster'

Fields

logEncryptionKmsKeyId :: Maybe Text

The KMS key used for encrypting log files. This attribute is only available with EMR version 5.30.0 and later, excluding EMR 6.0.0.

clusterArn :: Maybe Text

The Amazon Resource Name of the cluster.

requestedAmiVersion :: Maybe Text

The AMI version requested for this cluster.

ebsRootVolumeSize :: Maybe Int

The size, in GiB, of the Amazon EBS root device volume of the Linux AMI that is used for each EC2 instance. Available in Amazon EMR version 4.x and later.

ec2InstanceAttributes :: Maybe Ec2InstanceAttributes

Provides information about the EC2 instances in a cluster grouped by category. For example, key name, subnet ID, IAM instance profile, and so on.

outpostArn :: Maybe Text

The Amazon Resource Name (ARN) of the Outpost where the cluster is launched.

normalizedInstanceHours :: Maybe Int

An approximation of the cost of the cluster, represented in m1.small/hours. This value is incremented one time for every hour an m1.small instance runs. Larger instances are weighted more, so an EC2 instance that is roughly four times more expensive would result in the normalized instance hours being incremented by four. This result is only an approximation and does not reflect the actual billing rate.

configurations :: Maybe [Configuration]

Applies only to Amazon EMR releases 4.x and later. The list of Configurations supplied to the EMR cluster.

customAmiId :: Maybe Text

Available only in Amazon EMR version 5.7.0 and later. The ID of a custom Amazon EBS-backed Linux AMI if the cluster uses a custom AMI.

autoScalingRole :: Maybe Text

An IAM role for automatic scaling policies. The default role is EMR_AutoScaling_DefaultRole. The IAM role provides permissions that the automatic scaling feature requires to launch and terminate EC2 instances in an instance group.

securityConfiguration :: Maybe Text

The name of the security configuration applied to the cluster.

scaleDownBehavior :: Maybe ScaleDownBehavior

The way that individual Amazon EC2 instances terminate when an automatic scale-in activity occurs or an instance group is resized. TERMINATE_AT_INSTANCE_HOUR indicates that Amazon EMR terminates nodes at the instance-hour boundary, regardless of when the request to terminate the instance was submitted. This option is only available with Amazon EMR 5.1.0 and later and is the default for clusters created using that version. TERMINATE_AT_TASK_COMPLETION indicates that Amazon EMR adds nodes to a deny list and drains tasks from nodes before terminating the Amazon EC2 instances, regardless of the instance-hour boundary. With either behavior, Amazon EMR removes the least active nodes first and blocks instance termination if it could lead to HDFS corruption. TERMINATE_AT_TASK_COMPLETION is available only in Amazon EMR version 4.1.0 and later, and is the default for versions of Amazon EMR earlier than 5.1.0.

instanceCollectionType :: Maybe InstanceCollectionType

The instance fleet configuration is available only in Amazon EMR versions 4.8.0 and later, excluding 5.0.x versions.

The instance group configuration of the cluster. A value of INSTANCE_GROUP indicates a uniform instance group configuration. A value of INSTANCE_FLEET indicates an instance fleets configuration.

releaseLabel :: Maybe Text

The Amazon EMR release label, which determines the version of open-source application packages installed on the cluster. Release labels are in the form emr-x.x.x, where x.x.x is an Amazon EMR release version such as emr-5.14.0. For more information about Amazon EMR release versions and included application versions and features, see https://docs.aws.amazon.com/emr/latest/ReleaseGuide/. The release label applies only to Amazon EMR releases version 4.0 and later. Earlier versions use AmiVersion.

repoUpgradeOnBoot :: Maybe RepoUpgradeOnBoot

Applies only when CustomAmiID is used. Specifies the type of updates that are applied from the Amazon Linux AMI package repositories when an instance boots using the AMI.

logUri :: Maybe Text

The path to the Amazon S3 location where logs for this cluster are stored.

kerberosAttributes :: Maybe KerberosAttributes

Attributes for Kerberos configuration when Kerberos authentication is enabled using a security configuration. For more information see Use Kerberos Authentication in the Amazon EMR Management Guide.

placementGroups :: Maybe [PlacementGroupConfig]

Placement group configured for an Amazon EMR cluster.

runningAmiVersion :: Maybe Text

The AMI version running on this cluster.

masterPublicDnsName :: Maybe Text

The DNS name of the master node. If the cluster is on a private subnet, this is the private DNS name. On a public subnet, this is the public DNS name.

terminationProtected :: Maybe Bool

Indicates whether Amazon EMR will lock the cluster to prevent the EC2 instances from being terminated by an API call or user intervention, or in the event of a cluster error.

visibleToAllUsers :: Maybe Bool

Indicates whether the cluster is visible to IAM principals in the Amazon Web Services account associated with the cluster. When true, IAM principals in the Amazon Web Services account can perform EMR cluster actions on the cluster that their IAM policies allow. When false, only the IAM principal that created the cluster and the Amazon Web Services account root user can perform EMR actions, regardless of IAM permissions policies attached to other IAM principals.

The default value is true if a value is not provided when creating a cluster using the EMR API RunJobFlow command, the CLI create-cluster command, or the Amazon Web Services Management Console. IAM principals that are allowed to perform actions on the cluster can use the SetVisibleToAllUsers action to change the value on a running cluster. For more information, see Understanding the EMR Cluster VisibleToAllUsers Setting in the Amazon EMRManagement Guide.

autoTerminate :: Maybe Bool

Specifies whether the cluster should terminate after completing all steps.

stepConcurrencyLevel :: Maybe Int

Specifies the number of steps that can be executed concurrently.

applications :: Maybe [Application]

The applications installed on this cluster.

tags :: Maybe [Tag]

A list of tags associated with a cluster.

serviceRole :: Maybe Text

The IAM role that Amazon EMR assumes in order to access Amazon Web Services resources on your behalf.

id :: Text

The unique identifier for the cluster.

name :: Text

The name of the cluster.

status :: ClusterStatus

The current status details about the cluster.

newCluster

Arguments

:: Text
$sel:id:Cluster'

-> Text
$sel:name:Cluster'

-> ClusterStatus
$sel:status:Cluster'

-> Cluster

Create a value of Cluster with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:logEncryptionKmsKeyId:Cluster', cluster_logEncryptionKmsKeyId - The KMS key used for encrypting log files. This attribute is only available with EMR version 5.30.0 and later, excluding EMR 6.0.0.

$sel:clusterArn:Cluster', cluster_clusterArn - The Amazon Resource Name of the cluster.

$sel:requestedAmiVersion:Cluster', cluster_requestedAmiVersion - The AMI version requested for this cluster.

$sel:ebsRootVolumeSize:Cluster', cluster_ebsRootVolumeSize - The size, in GiB, of the Amazon EBS root device volume of the Linux AMI that is used for each EC2 instance. Available in Amazon EMR version 4.x and later.

$sel:ec2InstanceAttributes:Cluster', cluster_ec2InstanceAttributes - Provides information about the EC2 instances in a cluster grouped by category. For example, key name, subnet ID, IAM instance profile, and so on.

$sel:outpostArn:Cluster', cluster_outpostArn - The Amazon Resource Name (ARN) of the Outpost where the cluster is launched.

$sel:normalizedInstanceHours:Cluster', cluster_normalizedInstanceHours - An approximation of the cost of the cluster, represented in m1.small/hours. This value is incremented one time for every hour an m1.small instance runs. Larger instances are weighted more, so an EC2 instance that is roughly four times more expensive would result in the normalized instance hours being incremented by four. This result is only an approximation and does not reflect the actual billing rate.

$sel:configurations:Cluster', cluster_configurations - Applies only to Amazon EMR releases 4.x and later. The list of Configurations supplied to the EMR cluster.

$sel:customAmiId:Cluster', cluster_customAmiId - Available only in Amazon EMR version 5.7.0 and later. The ID of a custom Amazon EBS-backed Linux AMI if the cluster uses a custom AMI.

$sel:autoScalingRole:Cluster', cluster_autoScalingRole - An IAM role for automatic scaling policies. The default role is EMR_AutoScaling_DefaultRole. The IAM role provides permissions that the automatic scaling feature requires to launch and terminate EC2 instances in an instance group.

$sel:securityConfiguration:Cluster', cluster_securityConfiguration - The name of the security configuration applied to the cluster.

$sel:scaleDownBehavior:Cluster', cluster_scaleDownBehavior - The way that individual Amazon EC2 instances terminate when an automatic scale-in activity occurs or an instance group is resized. TERMINATE_AT_INSTANCE_HOUR indicates that Amazon EMR terminates nodes at the instance-hour boundary, regardless of when the request to terminate the instance was submitted. This option is only available with Amazon EMR 5.1.0 and later and is the default for clusters created using that version. TERMINATE_AT_TASK_COMPLETION indicates that Amazon EMR adds nodes to a deny list and drains tasks from nodes before terminating the Amazon EC2 instances, regardless of the instance-hour boundary. With either behavior, Amazon EMR removes the least active nodes first and blocks instance termination if it could lead to HDFS corruption. TERMINATE_AT_TASK_COMPLETION is available only in Amazon EMR version 4.1.0 and later, and is the default for versions of Amazon EMR earlier than 5.1.0.

$sel:instanceCollectionType:Cluster', cluster_instanceCollectionType - The instance fleet configuration is available only in Amazon EMR versions 4.8.0 and later, excluding 5.0.x versions.

The instance group configuration of the cluster. A value of INSTANCE_GROUP indicates a uniform instance group configuration. A value of INSTANCE_FLEET indicates an instance fleets configuration.

$sel:releaseLabel:Cluster', cluster_releaseLabel - The Amazon EMR release label, which determines the version of open-source application packages installed on the cluster. Release labels are in the form emr-x.x.x, where x.x.x is an Amazon EMR release version such as emr-5.14.0. For more information about Amazon EMR release versions and included application versions and features, see https://docs.aws.amazon.com/emr/latest/ReleaseGuide/. The release label applies only to Amazon EMR releases version 4.0 and later. Earlier versions use AmiVersion.

$sel:repoUpgradeOnBoot:Cluster', cluster_repoUpgradeOnBoot - Applies only when CustomAmiID is used. Specifies the type of updates that are applied from the Amazon Linux AMI package repositories when an instance boots using the AMI.

$sel:logUri:Cluster', cluster_logUri - The path to the Amazon S3 location where logs for this cluster are stored.

$sel:kerberosAttributes:Cluster', cluster_kerberosAttributes - Attributes for Kerberos configuration when Kerberos authentication is enabled using a security configuration. For more information see Use Kerberos Authentication in the Amazon EMR Management Guide.

$sel:placementGroups:Cluster', cluster_placementGroups - Placement group configured for an Amazon EMR cluster.

$sel:runningAmiVersion:Cluster', cluster_runningAmiVersion - The AMI version running on this cluster.

$sel:masterPublicDnsName:Cluster', cluster_masterPublicDnsName - The DNS name of the master node. If the cluster is on a private subnet, this is the private DNS name. On a public subnet, this is the public DNS name.

$sel:terminationProtected:Cluster', cluster_terminationProtected - Indicates whether Amazon EMR will lock the cluster to prevent the EC2 instances from being terminated by an API call or user intervention, or in the event of a cluster error.

$sel:visibleToAllUsers:Cluster', cluster_visibleToAllUsers - Indicates whether the cluster is visible to IAM principals in the Amazon Web Services account associated with the cluster. When true, IAM principals in the Amazon Web Services account can perform EMR cluster actions on the cluster that their IAM policies allow. When false, only the IAM principal that created the cluster and the Amazon Web Services account root user can perform EMR actions, regardless of IAM permissions policies attached to other IAM principals.

The default value is true if a value is not provided when creating a cluster using the EMR API RunJobFlow command, the CLI create-cluster command, or the Amazon Web Services Management Console. IAM principals that are allowed to perform actions on the cluster can use the SetVisibleToAllUsers action to change the value on a running cluster. For more information, see Understanding the EMR Cluster VisibleToAllUsers Setting in the Amazon EMRManagement Guide.

$sel:autoTerminate:Cluster', cluster_autoTerminate - Specifies whether the cluster should terminate after completing all steps.

$sel:stepConcurrencyLevel:Cluster', cluster_stepConcurrencyLevel - Specifies the number of steps that can be executed concurrently.

$sel:applications:Cluster', cluster_applications - The applications installed on this cluster.

$sel:tags:Cluster', cluster_tags - A list of tags associated with a cluster.

$sel:serviceRole:Cluster', cluster_serviceRole - The IAM role that Amazon EMR assumes in order to access Amazon Web Services resources on your behalf.

$sel:id:Cluster', cluster_id - The unique identifier for the cluster.

$sel:name:Cluster', cluster_name - The name of the cluster.

$sel:status:Cluster', cluster_status - The current status details about the cluster.

cluster_logEncryptionKmsKeyId :: Lens' Cluster (Maybe Text)

The KMS key used for encrypting log files. This attribute is only available with EMR version 5.30.0 and later, excluding EMR 6.0.0.

cluster_clusterArn :: Lens' Cluster (Maybe Text)

The Amazon Resource Name of the cluster.

cluster_requestedAmiVersion :: Lens' Cluster (Maybe Text)

The AMI version requested for this cluster.

cluster_ebsRootVolumeSize :: Lens' Cluster (Maybe Int)

The size, in GiB, of the Amazon EBS root device volume of the Linux AMI that is used for each EC2 instance. Available in Amazon EMR version 4.x and later.

cluster_ec2InstanceAttributes :: Lens' Cluster (Maybe Ec2InstanceAttributes)

Provides information about the EC2 instances in a cluster grouped by category. For example, key name, subnet ID, IAM instance profile, and so on.

cluster_outpostArn :: Lens' Cluster (Maybe Text)

The Amazon Resource Name (ARN) of the Outpost where the cluster is launched.

cluster_normalizedInstanceHours :: Lens' Cluster (Maybe Int)

An approximation of the cost of the cluster, represented in m1.small/hours. This value is incremented one time for every hour an m1.small instance runs. Larger instances are weighted more, so an EC2 instance that is roughly four times more expensive would result in the normalized instance hours being incremented by four. This result is only an approximation and does not reflect the actual billing rate.

cluster_configurations :: Lens' Cluster (Maybe [Configuration])

Applies only to Amazon EMR releases 4.x and later. The list of Configurations supplied to the EMR cluster.

cluster_customAmiId :: Lens' Cluster (Maybe Text)

Available only in Amazon EMR version 5.7.0 and later. The ID of a custom Amazon EBS-backed Linux AMI if the cluster uses a custom AMI.

cluster_autoScalingRole :: Lens' Cluster (Maybe Text)

An IAM role for automatic scaling policies. The default role is EMR_AutoScaling_DefaultRole. The IAM role provides permissions that the automatic scaling feature requires to launch and terminate EC2 instances in an instance group.

cluster_securityConfiguration :: Lens' Cluster (Maybe Text)

The name of the security configuration applied to the cluster.

cluster_scaleDownBehavior :: Lens' Cluster (Maybe ScaleDownBehavior)

The way that individual Amazon EC2 instances terminate when an automatic scale-in activity occurs or an instance group is resized. TERMINATE_AT_INSTANCE_HOUR indicates that Amazon EMR terminates nodes at the instance-hour boundary, regardless of when the request to terminate the instance was submitted. This option is only available with Amazon EMR 5.1.0 and later and is the default for clusters created using that version. TERMINATE_AT_TASK_COMPLETION indicates that Amazon EMR adds nodes to a deny list and drains tasks from nodes before terminating the Amazon EC2 instances, regardless of the instance-hour boundary. With either behavior, Amazon EMR removes the least active nodes first and blocks instance termination if it could lead to HDFS corruption. TERMINATE_AT_TASK_COMPLETION is available only in Amazon EMR version 4.1.0 and later, and is the default for versions of Amazon EMR earlier than 5.1.0.

cluster_instanceCollectionType :: Lens' Cluster (Maybe InstanceCollectionType)

The instance fleet configuration is available only in Amazon EMR versions 4.8.0 and later, excluding 5.0.x versions.

The instance group configuration of the cluster. A value of INSTANCE_GROUP indicates a uniform instance group configuration. A value of INSTANCE_FLEET indicates an instance fleets configuration.

cluster_releaseLabel :: Lens' Cluster (Maybe Text)

The Amazon EMR release label, which determines the version of open-source application packages installed on the cluster. Release labels are in the form emr-x.x.x, where x.x.x is an Amazon EMR release version such as emr-5.14.0. For more information about Amazon EMR release versions and included application versions and features, see https://docs.aws.amazon.com/emr/latest/ReleaseGuide/. The release label applies only to Amazon EMR releases version 4.0 and later. Earlier versions use AmiVersion.

cluster_repoUpgradeOnBoot :: Lens' Cluster (Maybe RepoUpgradeOnBoot)

Applies only when CustomAmiID is used. Specifies the type of updates that are applied from the Amazon Linux AMI package repositories when an instance boots using the AMI.

cluster_logUri :: Lens' Cluster (Maybe Text)

The path to the Amazon S3 location where logs for this cluster are stored.

cluster_kerberosAttributes :: Lens' Cluster (Maybe KerberosAttributes)

Attributes for Kerberos configuration when Kerberos authentication is enabled using a security configuration. For more information see Use Kerberos Authentication in the Amazon EMR Management Guide.

cluster_placementGroups :: Lens' Cluster (Maybe [PlacementGroupConfig])

Placement group configured for an Amazon EMR cluster.

cluster_runningAmiVersion :: Lens' Cluster (Maybe Text)

The AMI version running on this cluster.

cluster_masterPublicDnsName :: Lens' Cluster (Maybe Text)

The DNS name of the master node. If the cluster is on a private subnet, this is the private DNS name. On a public subnet, this is the public DNS name.

cluster_terminationProtected :: Lens' Cluster (Maybe Bool)

Indicates whether Amazon EMR will lock the cluster to prevent the EC2 instances from being terminated by an API call or user intervention, or in the event of a cluster error.

cluster_visibleToAllUsers :: Lens' Cluster (Maybe Bool)

Indicates whether the cluster is visible to IAM principals in the Amazon Web Services account associated with the cluster. When true, IAM principals in the Amazon Web Services account can perform EMR cluster actions on the cluster that their IAM policies allow. When false, only the IAM principal that created the cluster and the Amazon Web Services account root user can perform EMR actions, regardless of IAM permissions policies attached to other IAM principals.

The default value is true if a value is not provided when creating a cluster using the EMR API RunJobFlow command, the CLI create-cluster command, or the Amazon Web Services Management Console. IAM principals that are allowed to perform actions on the cluster can use the SetVisibleToAllUsers action to change the value on a running cluster. For more information, see Understanding the EMR Cluster VisibleToAllUsers Setting in the Amazon EMRManagement Guide.

cluster_autoTerminate :: Lens' Cluster (Maybe Bool)

Specifies whether the cluster should terminate after completing all steps.

cluster_stepConcurrencyLevel :: Lens' Cluster (Maybe Int)

Specifies the number of steps that can be executed concurrently.

cluster_applications :: Lens' Cluster (Maybe [Application])

The applications installed on this cluster.

cluster_tags :: Lens' Cluster (Maybe [Tag])

A list of tags associated with a cluster.

cluster_serviceRole :: Lens' Cluster (Maybe Text)

The IAM role that Amazon EMR assumes in order to access Amazon Web Services resources on your behalf.

cluster_id :: Lens' Cluster Text

The unique identifier for the cluster.

cluster_name :: Lens' Cluster Text

The name of the cluster.

cluster_status :: Lens' Cluster ClusterStatus

The current status details about the cluster.