Amazonka.EMR.Types.ClusterTimeline

Description

Documentation

data ClusterTimeline

Represents the timeline of the cluster's lifecycle.

See: newClusterTimeline smart constructor.

Constructors

ClusterTimeline'

Fields

readyDateTime :: Maybe POSIX

The date and time when the cluster was ready to run steps.

creationDateTime :: Maybe POSIX

The creation date and time of the cluster.

endDateTime :: Maybe POSIX

The date and time when the cluster was terminated.

newClusterTimeline :: ClusterTimeline

Create a value of ClusterTimeline with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:readyDateTime:ClusterTimeline', clusterTimeline_readyDateTime - The date and time when the cluster was ready to run steps.

$sel:creationDateTime:ClusterTimeline', clusterTimeline_creationDateTime - The creation date and time of the cluster.

$sel:endDateTime:ClusterTimeline', clusterTimeline_endDateTime - The date and time when the cluster was terminated.

clusterTimeline_readyDateTime :: Lens' ClusterTimeline (Maybe UTCTime)

The date and time when the cluster was ready to run steps.

clusterTimeline_creationDateTime :: Lens' ClusterTimeline (Maybe UTCTime)

The creation date and time of the cluster.

clusterTimeline_endDateTime :: Lens' ClusterTimeline (Maybe UTCTime)

The date and time when the cluster was terminated.