Amazonka.EMR.Types.Command

Description

Documentation

data Command

An entity describing an executable that runs on a cluster.

See: newCommand smart constructor.

Constructors

Command'

Fields

args :: Maybe [Text]

Arguments for Amazon EMR to pass to the command for execution.

scriptPath :: Maybe Text

The Amazon S3 location of the command script.

name :: Maybe Text

The name of the command.

newCommand :: Command

Create a value of Command with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:args:Command', command_args - Arguments for Amazon EMR to pass to the command for execution.

$sel:scriptPath:Command', command_scriptPath - The Amazon S3 location of the command script.

$sel:name:Command', command_name - The name of the command.

command_args :: Lens' Command (Maybe [Text])

Arguments for Amazon EMR to pass to the command for execution.

command_scriptPath :: Lens' Command (Maybe Text)

The Amazon S3 location of the command script.

command_name :: Lens' Command (Maybe Text)

The name of the command.