Amazonka.EMR.Types.EbsBlockDevice

Description

Documentation

data EbsBlockDevice

Configuration of requested EBS block device associated with the instance group.

See: newEbsBlockDevice smart constructor.

Constructors

EbsBlockDevice'

Fields

device :: Maybe Text

The device name that is exposed to the instance, such as /dev/sdh.

volumeSpecification :: Maybe VolumeSpecification

EBS volume specifications such as volume type, IOPS, and size (GiB) that will be requested for the EBS volume attached to an EC2 instance in the cluster.

newEbsBlockDevice :: EbsBlockDevice

Create a value of EbsBlockDevice with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:device:EbsBlockDevice', ebsBlockDevice_device - The device name that is exposed to the instance, such as /dev/sdh.

$sel:volumeSpecification:EbsBlockDevice', ebsBlockDevice_volumeSpecification - EBS volume specifications such as volume type, IOPS, and size (GiB) that will be requested for the EBS volume attached to an EC2 instance in the cluster.

ebsBlockDevice_device :: Lens' EbsBlockDevice (Maybe Text)

The device name that is exposed to the instance, such as /dev/sdh.

ebsBlockDevice_volumeSpecification :: Lens' EbsBlockDevice (Maybe VolumeSpecification)

EBS volume specifications such as volume type, IOPS, and size (GiB) that will be requested for the EBS volume attached to an EC2 instance in the cluster.