Amazonka.EMR.Types.ExecutionEngineConfig

Description

Documentation

data ExecutionEngineConfig

Specifies the execution engine (cluster) to run the notebook and perform the notebook execution, for example, an EMR cluster.

See: newExecutionEngineConfig smart constructor.

Constructors

ExecutionEngineConfig'

Fields

masterInstanceSecurityGroupId :: Maybe Text

An optional unique ID of an EC2 security group to associate with the master instance of the EMR cluster for this notebook execution. For more information see Specifying EC2 Security Groups for EMR Notebooks in the EMR Management Guide.

type' :: Maybe ExecutionEngineType

The type of execution engine. A value of EMR specifies an EMR cluster.

id :: Text

The unique identifier of the execution engine. For an EMR cluster, this is the cluster ID.

newExecutionEngineConfig

Arguments

:: Text
$sel:id:ExecutionEngineConfig'

-> ExecutionEngineConfig

Create a value of ExecutionEngineConfig with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:masterInstanceSecurityGroupId:ExecutionEngineConfig', executionEngineConfig_masterInstanceSecurityGroupId - An optional unique ID of an EC2 security group to associate with the master instance of the EMR cluster for this notebook execution. For more information see Specifying EC2 Security Groups for EMR Notebooks in the EMR Management Guide.

$sel:type':ExecutionEngineConfig', executionEngineConfig_type - The type of execution engine. A value of EMR specifies an EMR cluster.

$sel:id:ExecutionEngineConfig', executionEngineConfig_id - The unique identifier of the execution engine. For an EMR cluster, this is the cluster ID.

executionEngineConfig_masterInstanceSecurityGroupId :: Lens' ExecutionEngineConfig (Maybe Text)

An optional unique ID of an EC2 security group to associate with the master instance of the EMR cluster for this notebook execution. For more information see Specifying EC2 Security Groups for EMR Notebooks in the EMR Management Guide.

executionEngineConfig_type :: Lens' ExecutionEngineConfig (Maybe ExecutionEngineType)

The type of execution engine. A value of EMR specifies an EMR cluster.

executionEngineConfig_id :: Lens' ExecutionEngineConfig Text

The unique identifier of the execution engine. For an EMR cluster, this is the cluster ID.