Amazonka.EMR.Types.InstanceGroupModifyConfig

Description

Documentation

data InstanceGroupModifyConfig

Modify the size or configurations of an instance group.

See: newInstanceGroupModifyConfig smart constructor.

Constructors

InstanceGroupModifyConfig'

Fields

instanceCount :: Maybe Int

Target size for the instance group.

configurations :: Maybe [Configuration]

A list of new or modified configurations to apply for an instance group.

eC2InstanceIdsToTerminate :: Maybe [Text]

The EC2 InstanceIds to terminate. After you terminate the instances, the instance group will not return to its original requested size.

shrinkPolicy :: Maybe ShrinkPolicy

Policy for customizing shrink operations.

instanceGroupId :: Text

Unique ID of the instance group to modify.

newInstanceGroupModifyConfig

Arguments

:: Text
$sel:instanceGroupId:InstanceGroupModifyConfig'

-> InstanceGroupModifyConfig

Create a value of InstanceGroupModifyConfig with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:instanceCount:InstanceGroupModifyConfig', instanceGroupModifyConfig_instanceCount - Target size for the instance group.

$sel:configurations:InstanceGroupModifyConfig', instanceGroupModifyConfig_configurations - A list of new or modified configurations to apply for an instance group.

$sel:eC2InstanceIdsToTerminate:InstanceGroupModifyConfig', instanceGroupModifyConfig_eC2InstanceIdsToTerminate - The EC2 InstanceIds to terminate. After you terminate the instances, the instance group will not return to its original requested size.

$sel:shrinkPolicy:InstanceGroupModifyConfig', instanceGroupModifyConfig_shrinkPolicy - Policy for customizing shrink operations.

$sel:instanceGroupId:InstanceGroupModifyConfig', instanceGroupModifyConfig_instanceGroupId - Unique ID of the instance group to modify.

instanceGroupModifyConfig_instanceCount :: Lens' InstanceGroupModifyConfig (Maybe Int)

Target size for the instance group.

instanceGroupModifyConfig_configurations :: Lens' InstanceGroupModifyConfig (Maybe [Configuration])

A list of new or modified configurations to apply for an instance group.

instanceGroupModifyConfig_eC2InstanceIdsToTerminate :: Lens' InstanceGroupModifyConfig (Maybe [Text])

The EC2 InstanceIds to terminate. After you terminate the instances, the instance group will not return to its original requested size.

instanceGroupModifyConfig_shrinkPolicy :: Lens' InstanceGroupModifyConfig (Maybe ShrinkPolicy)

Policy for customizing shrink operations.

instanceGroupModifyConfig_instanceGroupId :: Lens' InstanceGroupModifyConfig Text

Unique ID of the instance group to modify.