Amazonka.EMR.Types.InstanceStatus

Description

Documentation

data InstanceStatus

The instance status details.

See: newInstanceStatus smart constructor.

Constructors

InstanceStatus'

Fields

state :: Maybe InstanceState

The current state of the instance.

stateChangeReason :: Maybe InstanceStateChangeReason

The details of the status change reason for the instance.

timeline :: Maybe InstanceTimeline

The timeline of the instance status over time.

newInstanceStatus :: InstanceStatus

Create a value of InstanceStatus with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:state:InstanceStatus', instanceStatus_state - The current state of the instance.

$sel:stateChangeReason:InstanceStatus', instanceStatus_stateChangeReason - The details of the status change reason for the instance.

$sel:timeline:InstanceStatus', instanceStatus_timeline - The timeline of the instance status over time.

instanceStatus_state :: Lens' InstanceStatus (Maybe InstanceState)

The current state of the instance.

instanceStatus_stateChangeReason :: Lens' InstanceStatus (Maybe InstanceStateChangeReason)

The details of the status change reason for the instance.

instanceStatus_timeline :: Lens' InstanceStatus (Maybe InstanceTimeline)

The timeline of the instance status over time.