Amazonka.EMR.Types.PlacementGroupConfig

Description

Documentation

data PlacementGroupConfig

Placement group configuration for an Amazon EMR cluster. The configuration specifies the placement strategy that can be applied to instance roles during cluster creation.

To use this configuration, consider attaching managed policy AmazonElasticMapReducePlacementGroupPolicy to the EMR role.

See: newPlacementGroupConfig smart constructor.

Constructors

PlacementGroupConfig'

Fields

placementStrategy :: Maybe PlacementGroupStrategy

EC2 Placement Group strategy associated with instance role.

Starting with Amazon EMR version 5.23.0, the only supported placement strategy is SPREAD for the MASTER instance role.

instanceRole :: InstanceRoleType

Role of the instance in the cluster.

Starting with Amazon EMR version 5.23.0, the only supported instance role is MASTER.

newPlacementGroupConfig

Arguments

:: InstanceRoleType
$sel:instanceRole:PlacementGroupConfig'

-> PlacementGroupConfig

Create a value of PlacementGroupConfig with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:placementStrategy:PlacementGroupConfig', placementGroupConfig_placementStrategy - EC2 Placement Group strategy associated with instance role.

Starting with Amazon EMR version 5.23.0, the only supported placement strategy is SPREAD for the MASTER instance role.

$sel:instanceRole:PlacementGroupConfig', placementGroupConfig_instanceRole - Role of the instance in the cluster.

Starting with Amazon EMR version 5.23.0, the only supported instance role is MASTER.

placementGroupConfig_placementStrategy :: Lens' PlacementGroupConfig (Maybe PlacementGroupStrategy)

EC2 Placement Group strategy associated with instance role.

Starting with Amazon EMR version 5.23.0, the only supported placement strategy is SPREAD for the MASTER instance role.

placementGroupConfig_instanceRole :: Lens' PlacementGroupConfig InstanceRoleType

Role of the instance in the cluster.

Starting with Amazon EMR version 5.23.0, the only supported instance role is MASTER.