Amazonka.EMR.Types.ScalingConstraints

Description

Documentation

data ScalingConstraints

The upper and lower EC2 instance limits for an automatic scaling policy. Automatic scaling activities triggered by automatic scaling rules will not cause an instance group to grow above or below these limits.

See: newScalingConstraints smart constructor.

Constructors

ScalingConstraints'

Fields

minCapacity :: Int

The lower boundary of EC2 instances in an instance group below which scaling activities are not allowed to shrink. Scale-in activities will not terminate instances below this boundary.

maxCapacity :: Int

The upper boundary of EC2 instances in an instance group beyond which scaling activities are not allowed to grow. Scale-out activities will not add instances beyond this boundary.

newScalingConstraints

Arguments

:: Int
$sel:minCapacity:ScalingConstraints'

-> Int
$sel:maxCapacity:ScalingConstraints'

-> ScalingConstraints

Create a value of ScalingConstraints with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:minCapacity:ScalingConstraints', scalingConstraints_minCapacity - The lower boundary of EC2 instances in an instance group below which scaling activities are not allowed to shrink. Scale-in activities will not terminate instances below this boundary.

$sel:maxCapacity:ScalingConstraints', scalingConstraints_maxCapacity - The upper boundary of EC2 instances in an instance group beyond which scaling activities are not allowed to grow. Scale-out activities will not add instances beyond this boundary.

scalingConstraints_minCapacity :: Lens' ScalingConstraints Int

The lower boundary of EC2 instances in an instance group below which scaling activities are not allowed to shrink. Scale-in activities will not terminate instances below this boundary.

scalingConstraints_maxCapacity :: Lens' ScalingConstraints Int

The upper boundary of EC2 instances in an instance group beyond which scaling activities are not allowed to grow. Scale-out activities will not add instances beyond this boundary.