Amazonka.EMR.Types.ScalingRule

Description

Documentation

data ScalingRule

A scale-in or scale-out rule that defines scaling activity, including the CloudWatch metric alarm that triggers activity, how EC2 instances are added or removed, and the periodicity of adjustments. The automatic scaling policy for an instance group can comprise one or more automatic scaling rules.

See: newScalingRule smart constructor.

Constructors

ScalingRule'

Fields

description :: Maybe Text

A friendly, more verbose description of the automatic scaling rule.

name :: Text

The name used to identify an automatic scaling rule. Rule names must be unique within a scaling policy.

action :: ScalingAction

The conditions that trigger an automatic scaling activity.

trigger :: ScalingTrigger

The CloudWatch alarm definition that determines when automatic scaling activity is triggered.

newScalingRule

Arguments

:: Text
$sel:name:ScalingRule'

-> ScalingAction
$sel:action:ScalingRule'

-> ScalingTrigger
$sel:trigger:ScalingRule'

-> ScalingRule

Create a value of ScalingRule with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:description:ScalingRule', scalingRule_description - A friendly, more verbose description of the automatic scaling rule.

$sel:name:ScalingRule', scalingRule_name - The name used to identify an automatic scaling rule. Rule names must be unique within a scaling policy.

$sel:action:ScalingRule', scalingRule_action - The conditions that trigger an automatic scaling activity.

$sel:trigger:ScalingRule', scalingRule_trigger - The CloudWatch alarm definition that determines when automatic scaling activity is triggered.

scalingRule_description :: Lens' ScalingRule (Maybe Text)

A friendly, more verbose description of the automatic scaling rule.

scalingRule_name :: Lens' ScalingRule Text

The name used to identify an automatic scaling rule. Rule names must be unique within a scaling policy.

scalingRule_action :: Lens' ScalingRule ScalingAction

The conditions that trigger an automatic scaling activity.

scalingRule_trigger :: Lens' ScalingRule ScalingTrigger

The CloudWatch alarm definition that determines when automatic scaling activity is triggered.