Amazonka.EMR.Types.SecurityConfigurationSummary

Description

Documentation

data SecurityConfigurationSummary

The creation date and time, and name, of a security configuration.

See: newSecurityConfigurationSummary smart constructor.

Constructors

SecurityConfigurationSummary'

Fields

name :: Maybe Text

The name of the security configuration.

creationDateTime :: Maybe POSIX

The date and time the security configuration was created.

newSecurityConfigurationSummary :: SecurityConfigurationSummary

Create a value of SecurityConfigurationSummary with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:name:SecurityConfigurationSummary', securityConfigurationSummary_name - The name of the security configuration.

$sel:creationDateTime:SecurityConfigurationSummary', securityConfigurationSummary_creationDateTime - The date and time the security configuration was created.

securityConfigurationSummary_name :: Lens' SecurityConfigurationSummary (Maybe Text)

The name of the security configuration.

securityConfigurationSummary_creationDateTime :: Lens' SecurityConfigurationSummary (Maybe UTCTime)

The date and time the security configuration was created.