Amazonka.EMR.Types.Studio

Description

Documentation

data Studio

Details for an Amazon EMR Studio including ID, creation time, name, and so on.

See: newStudio smart constructor.

Constructors

Studio'

Fields

creationTime :: Maybe POSIX

The time the Amazon EMR Studio was created.

engineSecurityGroupId :: Maybe Text

The ID of the Engine security group associated with the Amazon EMR Studio. The Engine security group allows inbound network traffic from resources in the Workspace security group.

subnetIds :: Maybe [Text]

The list of IDs of the subnets associated with the Amazon EMR Studio.

studioId :: Maybe Text

The ID of the Amazon EMR Studio.

vpcId :: Maybe Text

The ID of the VPC associated with the Amazon EMR Studio.

url :: Maybe Text

The unique access URL of the Amazon EMR Studio.

authMode :: Maybe AuthMode

Specifies whether the Amazon EMR Studio authenticates users using IAM or Amazon Web Services SSO.

defaultS3Location :: Maybe Text

The Amazon S3 location to back up Amazon EMR Studio Workspaces and notebook files.

idpAuthUrl :: Maybe Text

Your identity provider's authentication endpoint. Amazon EMR Studio redirects federated users to this endpoint for authentication when logging in to a Studio with the Studio URL.

workspaceSecurityGroupId :: Maybe Text

The ID of the Workspace security group associated with the Amazon EMR Studio. The Workspace security group allows outbound network traffic to resources in the Engine security group and to the internet.

name :: Maybe Text

The name of the Amazon EMR Studio.

idpRelayStateParameterName :: Maybe Text

The name of your identity provider's RelayState parameter.

studioArn :: Maybe Text

The Amazon Resource Name (ARN) of the Amazon EMR Studio.

userRole :: Maybe Text

The name of the IAM role assumed by users logged in to the Amazon EMR Studio. A Studio only requires a UserRole when you use IAM authentication.

description :: Maybe Text

The detailed description of the Amazon EMR Studio.

tags :: Maybe [Tag]

A list of tags associated with the Amazon EMR Studio.

serviceRole :: Maybe Text

The name of the IAM role assumed by the Amazon EMR Studio.

newStudio :: Studio

Create a value of Studio with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:creationTime:Studio', studio_creationTime - The time the Amazon EMR Studio was created.

$sel:engineSecurityGroupId:Studio', studio_engineSecurityGroupId - The ID of the Engine security group associated with the Amazon EMR Studio. The Engine security group allows inbound network traffic from resources in the Workspace security group.

$sel:subnetIds:Studio', studio_subnetIds - The list of IDs of the subnets associated with the Amazon EMR Studio.

$sel:studioId:Studio', studio_studioId - The ID of the Amazon EMR Studio.

$sel:vpcId:Studio', studio_vpcId - The ID of the VPC associated with the Amazon EMR Studio.

$sel:url:Studio', studio_url - The unique access URL of the Amazon EMR Studio.

$sel:authMode:Studio', studio_authMode - Specifies whether the Amazon EMR Studio authenticates users using IAM or Amazon Web Services SSO.

$sel:defaultS3Location:Studio', studio_defaultS3Location - The Amazon S3 location to back up Amazon EMR Studio Workspaces and notebook files.

$sel:idpAuthUrl:Studio', studio_idpAuthUrl - Your identity provider's authentication endpoint. Amazon EMR Studio redirects federated users to this endpoint for authentication when logging in to a Studio with the Studio URL.

$sel:workspaceSecurityGroupId:Studio', studio_workspaceSecurityGroupId - The ID of the Workspace security group associated with the Amazon EMR Studio. The Workspace security group allows outbound network traffic to resources in the Engine security group and to the internet.

$sel:name:Studio', studio_name - The name of the Amazon EMR Studio.

$sel:idpRelayStateParameterName:Studio', studio_idpRelayStateParameterName - The name of your identity provider's RelayState parameter.

$sel:studioArn:Studio', studio_studioArn - The Amazon Resource Name (ARN) of the Amazon EMR Studio.

$sel:userRole:Studio', studio_userRole - The name of the IAM role assumed by users logged in to the Amazon EMR Studio. A Studio only requires a UserRole when you use IAM authentication.

$sel:description:Studio', studio_description - The detailed description of the Amazon EMR Studio.

$sel:tags:Studio', studio_tags - A list of tags associated with the Amazon EMR Studio.

$sel:serviceRole:Studio', studio_serviceRole - The name of the IAM role assumed by the Amazon EMR Studio.

studio_creationTime :: Lens' Studio (Maybe UTCTime)

The time the Amazon EMR Studio was created.

studio_engineSecurityGroupId :: Lens' Studio (Maybe Text)

The ID of the Engine security group associated with the Amazon EMR Studio. The Engine security group allows inbound network traffic from resources in the Workspace security group.

studio_subnetIds :: Lens' Studio (Maybe [Text])

The list of IDs of the subnets associated with the Amazon EMR Studio.

studio_studioId :: Lens' Studio (Maybe Text)

The ID of the Amazon EMR Studio.

studio_vpcId :: Lens' Studio (Maybe Text)

The ID of the VPC associated with the Amazon EMR Studio.

studio_url :: Lens' Studio (Maybe Text)

The unique access URL of the Amazon EMR Studio.

studio_authMode :: Lens' Studio (Maybe AuthMode)

Specifies whether the Amazon EMR Studio authenticates users using IAM or Amazon Web Services SSO.

studio_defaultS3Location :: Lens' Studio (Maybe Text)

The Amazon S3 location to back up Amazon EMR Studio Workspaces and notebook files.

studio_idpAuthUrl :: Lens' Studio (Maybe Text)

Your identity provider's authentication endpoint. Amazon EMR Studio redirects federated users to this endpoint for authentication when logging in to a Studio with the Studio URL.

studio_workspaceSecurityGroupId :: Lens' Studio (Maybe Text)

The ID of the Workspace security group associated with the Amazon EMR Studio. The Workspace security group allows outbound network traffic to resources in the Engine security group and to the internet.

studio_name :: Lens' Studio (Maybe Text)

The name of the Amazon EMR Studio.

studio_idpRelayStateParameterName :: Lens' Studio (Maybe Text)

The name of your identity provider's RelayState parameter.

studio_studioArn :: Lens' Studio (Maybe Text)

The Amazon Resource Name (ARN) of the Amazon EMR Studio.

studio_userRole :: Lens' Studio (Maybe Text)

The name of the IAM role assumed by users logged in to the Amazon EMR Studio. A Studio only requires a UserRole when you use IAM authentication.

studio_description :: Lens' Studio (Maybe Text)

The detailed description of the Amazon EMR Studio.

studio_tags :: Lens' Studio (Maybe [Tag])

A list of tags associated with the Amazon EMR Studio.

studio_serviceRole :: Lens' Studio (Maybe Text)

The name of the IAM role assumed by the Amazon EMR Studio.