Amazonka.EMRContainers

Description

Derived from API version 2020-10-01 of the AWS service descriptions, licensed under Apache 2.0.

Amazon EMR on EKS provides a deployment option for Amazon EMR that allows you to run open-source big data frameworks on Amazon Elastic Kubernetes Service (Amazon EKS). With this deployment option, you can focus on running analytics workloads while Amazon EMR on EKS builds, configures, and manages containers for open-source applications. For more information about Amazon EMR on EKS concepts and tasks, see What is Amazon EMR on EKS.

Amazon EMR containers is the API name for Amazon EMR on EKS. The emr-containers prefix is used in the following scenarios:

It is the prefix in the CLI commands for Amazon EMR on EKS. For example, aws emr-containers start-job-run.

It is the prefix before IAM policy actions for Amazon EMR on EKS. For example, "Action": [ "emr-containers:StartJobRun"]. For more information, see Policy actions for Amazon EMR on EKS.

It is the prefix used in Amazon EMR on EKS service endpoints. For example, emr-containers.us-east-2.amazonaws.com. For more information, see Amazon EMR on EKS Service Endpoints.

Service Configuration

defaultService :: Service

API version 2020-10-01 of the Amazon EMR Containers SDK configuration.

Errors

Error matchers are designed for use with the functions provided by Control.Exception.Lens. This allows catching (and rethrowing) service specific errors returned by EMRContainers.

ValidationException

_ValidationException :: AsError a => Getting (First ServiceError) a ServiceError

There are invalid parameters in the client request.

InternalServerException

_InternalServerException :: AsError a => Getting (First ServiceError) a ServiceError

This is an internal server exception.

ResourceNotFoundException

_ResourceNotFoundException :: AsError a => Getting (First ServiceError) a ServiceError

The specified resource was not found.

Waiters

Waiters poll by repeatedly sending a request until some remote success condition configured by the Wait specification is fulfilled. The Wait specification determines how many attempts should be made, in addition to delay and retry strategies.

Operations

Some AWS operations return results that are incomplete and require subsequent requests in order to obtain the entire result set. The process of sending subsequent requests to continue where a previous request left off is called pagination. For example, the ListObjects operation of Amazon S3 returns up to 1000 objects at a time, and you must send subsequent requests with the appropriate Marker in order to retrieve the next page of results.

Operations that have an AWSPager instance can transparently perform subsequent requests, correctly setting Markers and other request facets to iterate through the entire result set of a truncated API operation. Operations which support this have an additional note in the documentation.

Many operations have the ability to filter results on the server side. See the individual operation parameters for details.

ListManagedEndpoints (Paginated)

data ListManagedEndpoints

See: newListManagedEndpoints smart constructor.

Constructors

ListManagedEndpoints' (Maybe [EndpointState]) (Maybe POSIX) (Maybe [Text]) (Maybe Text) (Maybe Int) (Maybe POSIX) Text

newListManagedEndpoints

Arguments

:: Text
$sel:virtualClusterId:ListManagedEndpoints'

-> ListManagedEndpoints

Create a value of ListManagedEndpoints with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:states:ListManagedEndpoints', listManagedEndpoints_states - The states of the managed endpoints.

$sel:createdAfter:ListManagedEndpoints', listManagedEndpoints_createdAfter - The date and time after which the endpoints are created.

$sel:types:ListManagedEndpoints', listManagedEndpoints_types - The types of the managed endpoints.

$sel:nextToken:ListManagedEndpoints', listManagedEndpoints_nextToken - The token for the next set of managed endpoints to return.

$sel:maxResults:ListManagedEndpoints', listManagedEndpoints_maxResults - The maximum number of managed endpoints that can be listed.

$sel:createdBefore:ListManagedEndpoints', listManagedEndpoints_createdBefore - The date and time before which the endpoints are created.

$sel:virtualClusterId:ListManagedEndpoints', listManagedEndpoints_virtualClusterId - The ID of the virtual cluster.

data ListManagedEndpointsResponse

See: newListManagedEndpointsResponse smart constructor.

Constructors

ListManagedEndpointsResponse' (Maybe Text) (Maybe [Endpoint]) Int

newListManagedEndpointsResponse

Arguments

:: Int
$sel:httpStatus:ListManagedEndpointsResponse'

-> ListManagedEndpointsResponse

Create a value of ListManagedEndpointsResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:nextToken:ListManagedEndpoints', listManagedEndpointsResponse_nextToken - The token for the next set of endpoints to return.

$sel:endpoints:ListManagedEndpointsResponse', listManagedEndpointsResponse_endpoints - The managed endpoints to be listed.

$sel:httpStatus:ListManagedEndpointsResponse', listManagedEndpointsResponse_httpStatus - The response's http status code.

CreateVirtualCluster

data CreateVirtualCluster

See: newCreateVirtualCluster smart constructor.

Constructors

CreateVirtualCluster' (Maybe (HashMap Text Text)) Text ContainerProvider Text

newCreateVirtualCluster

Arguments

:: Text
$sel:name:CreateVirtualCluster'

-> ContainerProvider
$sel:containerProvider:CreateVirtualCluster'

-> Text
$sel:clientToken:CreateVirtualCluster'

-> CreateVirtualCluster

Create a value of CreateVirtualCluster with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:tags:CreateVirtualCluster', createVirtualCluster_tags - The tags assigned to the virtual cluster.

$sel:name:CreateVirtualCluster', createVirtualCluster_name - The specified name of the virtual cluster.

$sel:containerProvider:CreateVirtualCluster', createVirtualCluster_containerProvider - The container provider of the virtual cluster.

$sel:clientToken:CreateVirtualCluster', createVirtualCluster_clientToken - The client token of the virtual cluster.

data CreateVirtualClusterResponse

See: newCreateVirtualClusterResponse smart constructor.

Constructors

CreateVirtualClusterResponse' (Maybe Text) (Maybe Text) (Maybe Text) Int

newCreateVirtualClusterResponse

Arguments

:: Int
$sel:httpStatus:CreateVirtualClusterResponse'

-> CreateVirtualClusterResponse

Create a value of CreateVirtualClusterResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:arn:CreateVirtualClusterResponse', createVirtualClusterResponse_arn - This output contains the ARN of virtual cluster.

$sel:name:CreateVirtualCluster', createVirtualClusterResponse_name - This output contains the name of the virtual cluster.

$sel:id:CreateVirtualClusterResponse', createVirtualClusterResponse_id - This output contains the virtual cluster ID.

$sel:httpStatus:CreateVirtualClusterResponse', createVirtualClusterResponse_httpStatus - The response's http status code.

DeleteVirtualCluster

data DeleteVirtualCluster

See: newDeleteVirtualCluster smart constructor.

Constructors

DeleteVirtualCluster' Text

newDeleteVirtualCluster

Arguments

:: Text
$sel:id:DeleteVirtualCluster'

-> DeleteVirtualCluster

Create a value of DeleteVirtualCluster with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:id:DeleteVirtualCluster', deleteVirtualCluster_id - The ID of the virtual cluster that will be deleted.

data DeleteVirtualClusterResponse

See: newDeleteVirtualClusterResponse smart constructor.

Constructors

DeleteVirtualClusterResponse' (Maybe Text) Int

newDeleteVirtualClusterResponse

Arguments

:: Int
$sel:httpStatus:DeleteVirtualClusterResponse'

-> DeleteVirtualClusterResponse

Create a value of DeleteVirtualClusterResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:id:DeleteVirtualCluster', deleteVirtualClusterResponse_id - This output contains the ID of the virtual cluster that will be deleted.

$sel:httpStatus:DeleteVirtualClusterResponse', deleteVirtualClusterResponse_httpStatus - The response's http status code.

CreateManagedEndpoint

data CreateManagedEndpoint

See: newCreateManagedEndpoint smart constructor.

Constructors

CreateManagedEndpoint' (Maybe ConfigurationOverrides) (Maybe (HashMap Text Text)) Text Text Text Text Text Text Text

newCreateManagedEndpoint

Arguments

:: Text
$sel:name:CreateManagedEndpoint'

-> Text
$sel:virtualClusterId:CreateManagedEndpoint'

-> Text
$sel:type':CreateManagedEndpoint'

-> Text
$sel:releaseLabel:CreateManagedEndpoint'

-> Text
$sel:executionRoleArn:CreateManagedEndpoint'

-> Text
$sel:certificateArn:CreateManagedEndpoint'

-> Text
$sel:clientToken:CreateManagedEndpoint'

-> CreateManagedEndpoint

Create a value of CreateManagedEndpoint with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:configurationOverrides:CreateManagedEndpoint', createManagedEndpoint_configurationOverrides - The configuration settings that will be used to override existing configurations.

$sel:tags:CreateManagedEndpoint', createManagedEndpoint_tags - The tags of the managed endpoint.

$sel:name:CreateManagedEndpoint', createManagedEndpoint_name - The name of the managed endpoint.

$sel:virtualClusterId:CreateManagedEndpoint', createManagedEndpoint_virtualClusterId - The ID of the virtual cluster for which a managed endpoint is created.

$sel:type':CreateManagedEndpoint', createManagedEndpoint_type - The type of the managed endpoint.

$sel:releaseLabel:CreateManagedEndpoint', createManagedEndpoint_releaseLabel - The Amazon EMR release version.

$sel:executionRoleArn:CreateManagedEndpoint', createManagedEndpoint_executionRoleArn - The ARN of the execution role.

$sel:certificateArn:CreateManagedEndpoint', createManagedEndpoint_certificateArn - The certificate ARN of the managed endpoint.

$sel:clientToken:CreateManagedEndpoint', createManagedEndpoint_clientToken - The client idempotency token for this create call.

data CreateManagedEndpointResponse

See: newCreateManagedEndpointResponse smart constructor.

Constructors

CreateManagedEndpointResponse' (Maybe Text) (Maybe Text) (Maybe Text) (Maybe Text) Int

newCreateManagedEndpointResponse

Arguments

:: Int
$sel:httpStatus:CreateManagedEndpointResponse'

-> CreateManagedEndpointResponse

Create a value of CreateManagedEndpointResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:arn:CreateManagedEndpointResponse', createManagedEndpointResponse_arn - The output contains the ARN of the managed endpoint.

$sel:name:CreateManagedEndpoint', createManagedEndpointResponse_name - The output contains the name of the managed endpoint.

$sel:id:CreateManagedEndpointResponse', createManagedEndpointResponse_id - The output contains the ID of the managed endpoint.

$sel:virtualClusterId:CreateManagedEndpoint', createManagedEndpointResponse_virtualClusterId - The output contains the ID of the virtual cluster.

$sel:httpStatus:CreateManagedEndpointResponse', createManagedEndpointResponse_httpStatus - The response's http status code.

ListTagsForResource

data ListTagsForResource

See: newListTagsForResource smart constructor.

Constructors

ListTagsForResource' Text

newListTagsForResource

Arguments

:: Text
$sel:resourceArn:ListTagsForResource'

-> ListTagsForResource

Create a value of ListTagsForResource with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:resourceArn:ListTagsForResource', listTagsForResource_resourceArn - The ARN of tagged resources.

data ListTagsForResourceResponse

See: newListTagsForResourceResponse smart constructor.

Constructors

ListTagsForResourceResponse' (Maybe (HashMap Text Text)) Int

newListTagsForResourceResponse

Arguments

:: Int
$sel:httpStatus:ListTagsForResourceResponse'

-> ListTagsForResourceResponse

Create a value of ListTagsForResourceResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:tags:ListTagsForResourceResponse', listTagsForResourceResponse_tags - The tags assigned to resources.

$sel:httpStatus:ListTagsForResourceResponse', listTagsForResourceResponse_httpStatus - The response's http status code.

CancelJobRun

data CancelJobRun

See: newCancelJobRun smart constructor.

Constructors

CancelJobRun' Text Text

newCancelJobRun

Arguments

:: Text
$sel:id:CancelJobRun'

-> Text
$sel:virtualClusterId:CancelJobRun'

-> CancelJobRun

Create a value of CancelJobRun with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:id:CancelJobRun', cancelJobRun_id - The ID of the job run to cancel.

$sel:virtualClusterId:CancelJobRun', cancelJobRun_virtualClusterId - The ID of the virtual cluster for which the job run will be canceled.

data CancelJobRunResponse

See: newCancelJobRunResponse smart constructor.

Constructors

CancelJobRunResponse' (Maybe Text) (Maybe Text) Int

newCancelJobRunResponse

Arguments

:: Int
$sel:httpStatus:CancelJobRunResponse'

-> CancelJobRunResponse

Create a value of CancelJobRunResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:id:CancelJobRun', cancelJobRunResponse_id - The output contains the ID of the cancelled job run.

$sel:virtualClusterId:CancelJobRun', cancelJobRunResponse_virtualClusterId - The output contains the virtual cluster ID for which the job run is cancelled.

$sel:httpStatus:CancelJobRunResponse', cancelJobRunResponse_httpStatus - The response's http status code.

DeleteManagedEndpoint

data DeleteManagedEndpoint

See: newDeleteManagedEndpoint smart constructor.

Constructors

DeleteManagedEndpoint' Text Text

newDeleteManagedEndpoint

Arguments

:: Text
$sel:id:DeleteManagedEndpoint'

-> Text
$sel:virtualClusterId:DeleteManagedEndpoint'

-> DeleteManagedEndpoint

Create a value of DeleteManagedEndpoint with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:id:DeleteManagedEndpoint', deleteManagedEndpoint_id - The ID of the managed endpoint.

$sel:virtualClusterId:DeleteManagedEndpoint', deleteManagedEndpoint_virtualClusterId - The ID of the endpoint's virtual cluster.

data DeleteManagedEndpointResponse

See: newDeleteManagedEndpointResponse smart constructor.

Constructors

DeleteManagedEndpointResponse' (Maybe Text) (Maybe Text) Int

newDeleteManagedEndpointResponse

Arguments

:: Int
$sel:httpStatus:DeleteManagedEndpointResponse'

-> DeleteManagedEndpointResponse

Create a value of DeleteManagedEndpointResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:id:DeleteManagedEndpoint', deleteManagedEndpointResponse_id - The output displays the ID of the managed endpoint.

$sel:virtualClusterId:DeleteManagedEndpoint', deleteManagedEndpointResponse_virtualClusterId - The output displays the ID of the endpoint's virtual cluster.

$sel:httpStatus:DeleteManagedEndpointResponse', deleteManagedEndpointResponse_httpStatus - The response's http status code.

ListJobRuns (Paginated)

data ListJobRuns

See: newListJobRuns smart constructor.

Constructors

ListJobRuns' (Maybe [JobRunState]) (Maybe POSIX) (Maybe Text) (Maybe Text) (Maybe Int) (Maybe POSIX) Text

newListJobRuns

Arguments

:: Text
$sel:virtualClusterId:ListJobRuns'

-> ListJobRuns

Create a value of ListJobRuns with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:states:ListJobRuns', listJobRuns_states - The states of the job run.

$sel:createdAfter:ListJobRuns', listJobRuns_createdAfter - The date and time after which the job runs were submitted.

$sel:nextToken:ListJobRuns', listJobRuns_nextToken - The token for the next set of job runs to return.

$sel:name:ListJobRuns', listJobRuns_name - The name of the job run.

$sel:maxResults:ListJobRuns', listJobRuns_maxResults - The maximum number of job runs that can be listed.

$sel:createdBefore:ListJobRuns', listJobRuns_createdBefore - The date and time before which the job runs were submitted.

$sel:virtualClusterId:ListJobRuns', listJobRuns_virtualClusterId - The ID of the virtual cluster for which to list the job run.

data ListJobRunsResponse

See: newListJobRunsResponse smart constructor.

Constructors

ListJobRunsResponse' (Maybe Text) (Maybe [JobRun]) Int

newListJobRunsResponse

Arguments

:: Int
$sel:httpStatus:ListJobRunsResponse'

-> ListJobRunsResponse

Create a value of ListJobRunsResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:nextToken:ListJobRuns', listJobRunsResponse_nextToken - This output displays the token for the next set of job runs.

$sel:jobRuns:ListJobRunsResponse', listJobRunsResponse_jobRuns - This output lists information about the specified job runs.

$sel:httpStatus:ListJobRunsResponse', listJobRunsResponse_httpStatus - The response's http status code.

ListVirtualClusters (Paginated)

data ListVirtualClusters

See: newListVirtualClusters smart constructor.

Constructors

ListVirtualClusters' (Maybe [VirtualClusterState]) (Maybe POSIX) (Maybe ContainerProviderType) (Maybe Text) (Maybe Text) (Maybe Int) (Maybe POSIX)

newListVirtualClusters :: ListVirtualClusters

Create a value of ListVirtualClusters with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:states:ListVirtualClusters', listVirtualClusters_states - The states of the requested virtual clusters.

$sel:createdAfter:ListVirtualClusters', listVirtualClusters_createdAfter - The date and time after which the virtual clusters are created.

$sel:containerProviderType:ListVirtualClusters', listVirtualClusters_containerProviderType - The container provider type of the virtual cluster. EKS is the only supported type as of now.

$sel:nextToken:ListVirtualClusters', listVirtualClusters_nextToken - The token for the next set of virtual clusters to return.

$sel:containerProviderId:ListVirtualClusters', listVirtualClusters_containerProviderId - The container provider ID of the virtual cluster.

$sel:maxResults:ListVirtualClusters', listVirtualClusters_maxResults - The maximum number of virtual clusters that can be listed.

$sel:createdBefore:ListVirtualClusters', listVirtualClusters_createdBefore - The date and time before which the virtual clusters are created.

data ListVirtualClustersResponse

See: newListVirtualClustersResponse smart constructor.

Constructors

ListVirtualClustersResponse' (Maybe Text) (Maybe [VirtualCluster]) Int

newListVirtualClustersResponse

Arguments

:: Int
$sel:httpStatus:ListVirtualClustersResponse'

-> ListVirtualClustersResponse

Create a value of ListVirtualClustersResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:nextToken:ListVirtualClusters', listVirtualClustersResponse_nextToken - This output displays the token for the next set of virtual clusters.

$sel:virtualClusters:ListVirtualClustersResponse', listVirtualClustersResponse_virtualClusters - This output lists the specified virtual clusters.

$sel:httpStatus:ListVirtualClustersResponse', listVirtualClustersResponse_httpStatus - The response's http status code.

TagResource

data TagResource

See: newTagResource smart constructor.

Constructors

TagResource' Text (HashMap Text Text)

newTagResource

Arguments

:: Text
$sel:resourceArn:TagResource'

-> TagResource

Create a value of TagResource with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:resourceArn:TagResource', tagResource_resourceArn - The ARN of resources.

$sel:tags:TagResource', tagResource_tags - The tags assigned to resources.

data TagResourceResponse

See: newTagResourceResponse smart constructor.

Constructors

TagResourceResponse' Int

newTagResourceResponse

Arguments

:: Int
$sel:httpStatus:TagResourceResponse'

-> TagResourceResponse

Create a value of TagResourceResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:httpStatus:TagResourceResponse', tagResourceResponse_httpStatus - The response's http status code.

DescribeManagedEndpoint

data DescribeManagedEndpoint

See: newDescribeManagedEndpoint smart constructor.

Constructors

DescribeManagedEndpoint' Text Text

newDescribeManagedEndpoint

Arguments

:: Text
$sel:id:DescribeManagedEndpoint'

-> Text
$sel:virtualClusterId:DescribeManagedEndpoint'

-> DescribeManagedEndpoint

Create a value of DescribeManagedEndpoint with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:id:DescribeManagedEndpoint', describeManagedEndpoint_id - This output displays ID of the managed endpoint.

$sel:virtualClusterId:DescribeManagedEndpoint', describeManagedEndpoint_virtualClusterId - The ID of the endpoint's virtual cluster.

data DescribeManagedEndpointResponse

See: newDescribeManagedEndpointResponse smart constructor.

Constructors

DescribeManagedEndpointResponse' (Maybe Endpoint) Int

newDescribeManagedEndpointResponse

Arguments

:: Int
$sel:httpStatus:DescribeManagedEndpointResponse'

-> DescribeManagedEndpointResponse

Create a value of DescribeManagedEndpointResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:endpoint:DescribeManagedEndpointResponse', describeManagedEndpointResponse_endpoint - This output displays information about a managed endpoint.

$sel:httpStatus:DescribeManagedEndpointResponse', describeManagedEndpointResponse_httpStatus - The response's http status code.

DescribeJobRun

data DescribeJobRun

See: newDescribeJobRun smart constructor.

Constructors

DescribeJobRun' Text Text

newDescribeJobRun

Arguments

:: Text
$sel:id:DescribeJobRun'

-> Text
$sel:virtualClusterId:DescribeJobRun'

-> DescribeJobRun

Create a value of DescribeJobRun with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:id:DescribeJobRun', describeJobRun_id - The ID of the job run request.

$sel:virtualClusterId:DescribeJobRun', describeJobRun_virtualClusterId - The ID of the virtual cluster for which the job run is submitted.

data DescribeJobRunResponse

See: newDescribeJobRunResponse smart constructor.

Constructors

DescribeJobRunResponse' (Maybe JobRun) Int

newDescribeJobRunResponse

Arguments

:: Int
$sel:httpStatus:DescribeJobRunResponse'

-> DescribeJobRunResponse

Create a value of DescribeJobRunResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:jobRun:DescribeJobRunResponse', describeJobRunResponse_jobRun - The output displays information about a job run.

$sel:httpStatus:DescribeJobRunResponse', describeJobRunResponse_httpStatus - The response's http status code.

UntagResource

data UntagResource

See: newUntagResource smart constructor.

Constructors

UntagResource' Text [Text]

newUntagResource

Arguments

:: Text
$sel:resourceArn:UntagResource'

-> UntagResource

Create a value of UntagResource with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:resourceArn:UntagResource', untagResource_resourceArn - The ARN of resources.

$sel:tagKeys:UntagResource', untagResource_tagKeys - The tag keys of the resources.

data UntagResourceResponse

See: newUntagResourceResponse smart constructor.

Constructors

UntagResourceResponse' Int

newUntagResourceResponse

Arguments

:: Int
$sel:httpStatus:UntagResourceResponse'

-> UntagResourceResponse

Create a value of UntagResourceResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:httpStatus:UntagResourceResponse', untagResourceResponse_httpStatus - The response's http status code.

DescribeVirtualCluster

data DescribeVirtualCluster

See: newDescribeVirtualCluster smart constructor.

Constructors

DescribeVirtualCluster' Text

newDescribeVirtualCluster

Arguments

:: Text
$sel:id:DescribeVirtualCluster'

-> DescribeVirtualCluster

Create a value of DescribeVirtualCluster with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:id:DescribeVirtualCluster', describeVirtualCluster_id - The ID of the virtual cluster that will be described.

data DescribeVirtualClusterResponse

See: newDescribeVirtualClusterResponse smart constructor.

Constructors

DescribeVirtualClusterResponse' (Maybe VirtualCluster) Int

newDescribeVirtualClusterResponse

Arguments

:: Int
$sel:httpStatus:DescribeVirtualClusterResponse'

-> DescribeVirtualClusterResponse

Create a value of DescribeVirtualClusterResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:virtualCluster:DescribeVirtualClusterResponse', describeVirtualClusterResponse_virtualCluster - This output displays information about the specified virtual cluster.

$sel:httpStatus:DescribeVirtualClusterResponse', describeVirtualClusterResponse_httpStatus - The response's http status code.

StartJobRun

data StartJobRun

See: newStartJobRun smart constructor.

Constructors

StartJobRun' (Maybe ConfigurationOverrides) (Maybe Text) (Maybe (HashMap Text Text)) Text Text Text Text JobDriver

newStartJobRun

Arguments

:: Text
$sel:virtualClusterId:StartJobRun'

-> Text
$sel:clientToken:StartJobRun'

-> Text
$sel:executionRoleArn:StartJobRun'

-> Text
$sel:releaseLabel:StartJobRun'

-> JobDriver
$sel:jobDriver:StartJobRun'

-> StartJobRun

Create a value of StartJobRun with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:configurationOverrides:StartJobRun', startJobRun_configurationOverrides - The configuration overrides for the job run.

$sel:name:StartJobRun', startJobRun_name - The name of the job run.

$sel:tags:StartJobRun', startJobRun_tags - The tags assigned to job runs.

$sel:virtualClusterId:StartJobRun', startJobRun_virtualClusterId - The virtual cluster ID for which the job run request is submitted.

$sel:clientToken:StartJobRun', startJobRun_clientToken - The client idempotency token of the job run request.

$sel:executionRoleArn:StartJobRun', startJobRun_executionRoleArn - The execution role ARN for the job run.

$sel:releaseLabel:StartJobRun', startJobRun_releaseLabel - The Amazon EMR release version to use for the job run.

$sel:jobDriver:StartJobRun', startJobRun_jobDriver - The job driver for the job run.

data StartJobRunResponse

See: newStartJobRunResponse smart constructor.

Constructors

StartJobRunResponse' (Maybe Text) (Maybe Text) (Maybe Text) (Maybe Text) Int

newStartJobRunResponse

Arguments

:: Int
$sel:httpStatus:StartJobRunResponse'

-> StartJobRunResponse

Create a value of StartJobRunResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:arn:StartJobRunResponse', startJobRunResponse_arn - This output lists the ARN of job run.

$sel:name:StartJobRun', startJobRunResponse_name - This output displays the name of the started job run.

$sel:id:StartJobRunResponse', startJobRunResponse_id - This output displays the started job run ID.

$sel:virtualClusterId:StartJobRun', startJobRunResponse_virtualClusterId - This output displays the virtual cluster ID for which the job run was submitted.

$sel:httpStatus:StartJobRunResponse', startJobRunResponse_httpStatus - The response's http status code.

Types

ContainerProviderType

newtype ContainerProviderType

Constructors

ContainerProviderType'

Fields

fromContainerProviderType :: Text

Bundled Patterns

pattern ContainerProviderType_EKS :: ContainerProviderType

EndpointState

newtype EndpointState

Constructors

EndpointState'

Fields

fromEndpointState :: Text

Bundled Patterns

pattern EndpointState_ACTIVE :: EndpointState

pattern EndpointState_CREATING :: EndpointState

pattern EndpointState_TERMINATED :: EndpointState

pattern EndpointState_TERMINATED_WITH_ERRORS :: EndpointState

pattern EndpointState_TERMINATING :: EndpointState

FailureReason

newtype FailureReason

Constructors

FailureReason'

Fields

fromFailureReason :: Text

Bundled Patterns

pattern FailureReason_CLUSTER_UNAVAILABLE :: FailureReason

pattern FailureReason_INTERNAL_ERROR :: FailureReason

pattern FailureReason_USER_ERROR :: FailureReason

pattern FailureReason_VALIDATION_ERROR :: FailureReason

JobRunState

newtype JobRunState

Constructors

JobRunState'

Fields

fromJobRunState :: Text

Bundled Patterns

pattern JobRunState_CANCELLED :: JobRunState

pattern JobRunState_CANCEL_PENDING :: JobRunState

pattern JobRunState_COMPLETED :: JobRunState

pattern JobRunState_FAILED :: JobRunState

pattern JobRunState_PENDING :: JobRunState

pattern JobRunState_RUNNING :: JobRunState

pattern JobRunState_SUBMITTED :: JobRunState

PersistentAppUI

newtype PersistentAppUI

Constructors

PersistentAppUI'

Fields

fromPersistentAppUI :: Text

Bundled Patterns

pattern PersistentAppUI_DISABLED :: PersistentAppUI

pattern PersistentAppUI_ENABLED :: PersistentAppUI

VirtualClusterState

newtype VirtualClusterState

Constructors

VirtualClusterState'

Fields

fromVirtualClusterState :: Text

Bundled Patterns

pattern VirtualClusterState_ARRESTED :: VirtualClusterState

pattern VirtualClusterState_RUNNING :: VirtualClusterState

pattern VirtualClusterState_TERMINATED :: VirtualClusterState

pattern VirtualClusterState_TERMINATING :: VirtualClusterState

CloudWatchMonitoringConfiguration

data CloudWatchMonitoringConfiguration

A configuration for CloudWatch monitoring. You can configure your jobs to send log information to CloudWatch Logs.

See: newCloudWatchMonitoringConfiguration smart constructor.

Constructors

CloudWatchMonitoringConfiguration' (Maybe Text) Text

newCloudWatchMonitoringConfiguration

Arguments

:: Text
$sel:logGroupName:CloudWatchMonitoringConfiguration'

-> CloudWatchMonitoringConfiguration

Create a value of CloudWatchMonitoringConfiguration with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:logStreamNamePrefix:CloudWatchMonitoringConfiguration', cloudWatchMonitoringConfiguration_logStreamNamePrefix - The specified name prefix for log streams.

$sel:logGroupName:CloudWatchMonitoringConfiguration', cloudWatchMonitoringConfiguration_logGroupName - The name of the log group for log publishing.

Configuration

data Configuration

A configuration specification to be used when provisioning virtual clusters, which can include configurations for applications and software bundled with Amazon EMR on EKS. A configuration consists of a classification, properties, and optional nested configurations. A classification refers to an application-specific configuration file. Properties are the settings you want to change in that file.

See: newConfiguration smart constructor.

Constructors

Configuration' (Maybe [Configuration]) (Maybe (Sensitive (HashMap Text Text))) Text

newConfiguration

Arguments

:: Text
$sel:classification:Configuration'

-> Configuration

Create a value of Configuration with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:configurations:Configuration', configuration_configurations - A list of additional configurations to apply within a configuration object.

$sel:properties:Configuration', configuration_properties - A set of properties specified within a configuration classification.

$sel:classification:Configuration', configuration_classification - The classification within a configuration.

ConfigurationOverrides

data ConfigurationOverrides

A configuration specification to be used to override existing configurations.

See: newConfigurationOverrides smart constructor.

Constructors

ConfigurationOverrides' (Maybe MonitoringConfiguration) (Maybe [Configuration])

newConfigurationOverrides :: ConfigurationOverrides

Create a value of ConfigurationOverrides with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:monitoringConfiguration:ConfigurationOverrides', configurationOverrides_monitoringConfiguration - The configurations for monitoring.

$sel:applicationConfiguration:ConfigurationOverrides', configurationOverrides_applicationConfiguration - The configurations for the application running by the job run.

ContainerInfo

data ContainerInfo

The information about the container used for a job run or a managed endpoint.

See: newContainerInfo smart constructor.

Constructors

ContainerInfo' (Maybe EksInfo)

newContainerInfo :: ContainerInfo

Create a value of ContainerInfo with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:eksInfo:ContainerInfo', containerInfo_eksInfo - The information about the EKS cluster.

ContainerProvider

data ContainerProvider

The information about the container provider.

See: newContainerProvider smart constructor.

Constructors

ContainerProvider' (Maybe ContainerInfo) ContainerProviderType Text

newContainerProvider

Arguments

:: ContainerProviderType
$sel:type':ContainerProvider'

-> Text
$sel:id:ContainerProvider'

-> ContainerProvider

Create a value of ContainerProvider with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:info:ContainerProvider', containerProvider_info - The information about the container cluster.

$sel:type':ContainerProvider', containerProvider_type - The type of the container provider. EKS is the only supported type as of now.

$sel:id:ContainerProvider', containerProvider_id - The ID of the container cluster.

EksInfo

data EksInfo

The information about the EKS cluster.

See: newEksInfo smart constructor.

Constructors

EksInfo' (Maybe Text)

newEksInfo :: EksInfo

Create a value of EksInfo with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:namespace:EksInfo', eksInfo_namespace - The namespaces of the EKS cluster.

Endpoint

data Endpoint

This entity represents the endpoint that is managed by Amazon EMR on EKS.

See: newEndpoint smart constructor.

Constructors

Endpoint' (Maybe FailureReason) (Maybe EndpointState) (Maybe Text) (Maybe POSIX) (Maybe [Text]) (Maybe Text) (Maybe Text) (Maybe Text) (Maybe Text) (Maybe ConfigurationOverrides) (Maybe Text) (Maybe Text) (Maybe Text) (Maybe Text) (Maybe Text) (Maybe Text) (Maybe (HashMap Text Text))

newEndpoint :: Endpoint

Create a value of Endpoint with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:failureReason:Endpoint', endpoint_failureReason - The reasons why the endpoint has failed.

$sel:state:Endpoint', endpoint_state - The state of the endpoint.

$sel:arn:Endpoint', endpoint_arn - The ARN of the endpoint.

$sel:createdAt:Endpoint', endpoint_createdAt - The date and time when the endpoint was created.

$sel:subnetIds:Endpoint', endpoint_subnetIds - The subnet IDs of the endpoint.

$sel:stateDetails:Endpoint', endpoint_stateDetails - Additional details of the endpoint state.

$sel:certificateArn:Endpoint', endpoint_certificateArn - The certificate ARN of the endpoint.

$sel:executionRoleArn:Endpoint', endpoint_executionRoleArn - The execution role ARN of the endpoint.

$sel:securityGroup:Endpoint', endpoint_securityGroup - The security group configuration of the endpoint.

$sel:configurationOverrides:Endpoint', endpoint_configurationOverrides - The configuration settings that are used to override existing configurations for endpoints.

$sel:name:Endpoint', endpoint_name - The name of the endpoint.

$sel:releaseLabel:Endpoint', endpoint_releaseLabel - The EMR release version to be used for the endpoint.

$sel:id:Endpoint', endpoint_id - The ID of the endpoint.

$sel:type':Endpoint', endpoint_type - The type of the endpoint.

$sel:serverUrl:Endpoint', endpoint_serverUrl - The server URL of the endpoint.

$sel:virtualClusterId:Endpoint', endpoint_virtualClusterId - The ID of the endpoint's virtual cluster.

$sel:tags:Endpoint', endpoint_tags - The tags of the endpoint.

JobDriver

data JobDriver

Specify the driver that the job runs on.

See: newJobDriver smart constructor.

Constructors

JobDriver' (Maybe SparkSubmitJobDriver)

newJobDriver :: JobDriver

Create a value of JobDriver with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:sparkSubmitJobDriver:JobDriver', jobDriver_sparkSubmitJobDriver - The job driver parameters specified for spark submit.

JobRun

data JobRun

This entity describes a job run. A job run is a unit of work, such as a Spark jar, PySpark script, or SparkSQL query, that you submit to Amazon EMR on EKS.

See: newJobRun smart constructor.

Constructors

JobRun' (Maybe FailureReason) (Maybe JobRunState) (Maybe Text) (Maybe Text) (Maybe POSIX) (Maybe Text) (Maybe Text) (Maybe Text) (Maybe JobDriver) (Maybe ConfigurationOverrides) (Maybe POSIX) (Maybe Text) (Maybe Text) (Maybe Text) (Maybe Text) (Maybe (HashMap Text Text))

newJobRun :: JobRun

Create a value of JobRun with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:failureReason:JobRun', jobRun_failureReason - The reasons why the job run has failed.

$sel:state:JobRun', jobRun_state - The state of the job run.

$sel:clientToken:JobRun', jobRun_clientToken - The client token used to start a job run.

$sel:arn:JobRun', jobRun_arn - The ARN of job run.

$sel:createdAt:JobRun', jobRun_createdAt - The date and time when the job run was created.

$sel:stateDetails:JobRun', jobRun_stateDetails - Additional details of the job run state.

$sel:createdBy:JobRun', jobRun_createdBy - The user who created the job run.

$sel:executionRoleArn:JobRun', jobRun_executionRoleArn - The execution role ARN of the job run.

$sel:jobDriver:JobRun', jobRun_jobDriver - Parameters of job driver for the job run.

$sel:configurationOverrides:JobRun', jobRun_configurationOverrides - The configuration settings that are used to override default configuration.

$sel:finishedAt:JobRun', jobRun_finishedAt - The date and time when the job run has finished.

$sel:name:JobRun', jobRun_name - The name of the job run.

$sel:releaseLabel:JobRun', jobRun_releaseLabel - The release version of Amazon EMR.

$sel:id:JobRun', jobRun_id - The ID of the job run.

$sel:virtualClusterId:JobRun', jobRun_virtualClusterId - The ID of the job run's virtual cluster.

$sel:tags:JobRun', jobRun_tags - The assigned tags of the job run.

MonitoringConfiguration

data MonitoringConfiguration

Configuration setting for monitoring.

See: newMonitoringConfiguration smart constructor.

Constructors

MonitoringConfiguration' (Maybe PersistentAppUI) (Maybe S3MonitoringConfiguration) (Maybe CloudWatchMonitoringConfiguration)

newMonitoringConfiguration :: MonitoringConfiguration

Create a value of MonitoringConfiguration with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:persistentAppUI:MonitoringConfiguration', monitoringConfiguration_persistentAppUI - Monitoring configurations for the persistent application UI.

$sel:s3MonitoringConfiguration:MonitoringConfiguration', monitoringConfiguration_s3MonitoringConfiguration - Amazon S3 configuration for monitoring log publishing.

$sel:cloudWatchMonitoringConfiguration:MonitoringConfiguration', monitoringConfiguration_cloudWatchMonitoringConfiguration - Monitoring configurations for CloudWatch.

S3MonitoringConfiguration

data S3MonitoringConfiguration

Amazon S3 configuration for monitoring log publishing. You can configure your jobs to send log information to Amazon S3.

See: newS3MonitoringConfiguration smart constructor.

Constructors

S3MonitoringConfiguration' Text

newS3MonitoringConfiguration

Arguments

:: Text
$sel:logUri:S3MonitoringConfiguration'

-> S3MonitoringConfiguration

Create a value of S3MonitoringConfiguration with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:logUri:S3MonitoringConfiguration', s3MonitoringConfiguration_logUri - Amazon S3 destination URI for log publishing.

SparkSubmitJobDriver

data SparkSubmitJobDriver

The information about job driver for Spark submit.

See: newSparkSubmitJobDriver smart constructor.

Constructors

SparkSubmitJobDriver' (Maybe (Sensitive Text)) (Maybe [Sensitive Text]) (Sensitive Text)

newSparkSubmitJobDriver

Arguments

:: Text
$sel:entryPoint:SparkSubmitJobDriver'

-> SparkSubmitJobDriver

Create a value of SparkSubmitJobDriver with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:sparkSubmitParameters:SparkSubmitJobDriver', sparkSubmitJobDriver_sparkSubmitParameters - The Spark submit parameters that are used for job runs.

$sel:entryPointArguments:SparkSubmitJobDriver', sparkSubmitJobDriver_entryPointArguments - The arguments for job application.

$sel:entryPoint:SparkSubmitJobDriver', sparkSubmitJobDriver_entryPoint - The entry point of job application.

VirtualCluster

data VirtualCluster

This entity describes a virtual cluster. A virtual cluster is a Kubernetes namespace that Amazon EMR is registered with. Amazon EMR uses virtual clusters to run jobs and host endpoints. Multiple virtual clusters can be backed by the same physical cluster. However, each virtual cluster maps to one namespace on an EKS cluster. Virtual clusters do not create any active resources that contribute to your bill or that require lifecycle management outside the service.

See: newVirtualCluster smart constructor.

Constructors

VirtualCluster' (Maybe VirtualClusterState) (Maybe Text) (Maybe POSIX) (Maybe Text) (Maybe Text) (Maybe ContainerProvider) (Maybe (HashMap Text Text))

newVirtualCluster :: VirtualCluster

Create a value of VirtualCluster with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:state:VirtualCluster', virtualCluster_state - The state of the virtual cluster.

$sel:arn:VirtualCluster', virtualCluster_arn - The ARN of the virtual cluster.

$sel:createdAt:VirtualCluster', virtualCluster_createdAt - The date and time when the virtual cluster is created.

$sel:name:VirtualCluster', virtualCluster_name - The name of the virtual cluster.

$sel:id:VirtualCluster', virtualCluster_id - The ID of the virtual cluster.

$sel:containerProvider:VirtualCluster', virtualCluster_containerProvider - The container provider of the virtual cluster.

$sel:tags:VirtualCluster', virtualCluster_tags - The assigned tags of the virtual cluster.