Amazonka.EMRContainers.CreateManagedEndpoint

Description

Creates a managed endpoint. A managed endpoint is a gateway that connects EMR Studio to Amazon EMR on EKS so that EMR Studio can communicate with your virtual cluster.

Creating a Request

data CreateManagedEndpoint

See: newCreateManagedEndpoint smart constructor.

Constructors

CreateManagedEndpoint'

Fields

configurationOverrides :: Maybe ConfigurationOverrides

The configuration settings that will be used to override existing configurations.

tags :: Maybe (HashMap Text Text)

The tags of the managed endpoint.

name :: Text

The name of the managed endpoint.

virtualClusterId :: Text

The ID of the virtual cluster for which a managed endpoint is created.

type' :: Text

The type of the managed endpoint.

releaseLabel :: Text

The Amazon EMR release version.

executionRoleArn :: Text

The ARN of the execution role.

certificateArn :: Text

The certificate ARN of the managed endpoint.

clientToken :: Text

The client idempotency token for this create call.

newCreateManagedEndpoint

Arguments

:: Text
$sel:name:CreateManagedEndpoint'

-> Text
$sel:virtualClusterId:CreateManagedEndpoint'

-> Text
$sel:type':CreateManagedEndpoint'

-> Text
$sel:releaseLabel:CreateManagedEndpoint'

-> Text
$sel:executionRoleArn:CreateManagedEndpoint'

-> Text
$sel:certificateArn:CreateManagedEndpoint'

-> Text
$sel:clientToken:CreateManagedEndpoint'

-> CreateManagedEndpoint

Create a value of CreateManagedEndpoint with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:configurationOverrides:CreateManagedEndpoint', createManagedEndpoint_configurationOverrides - The configuration settings that will be used to override existing configurations.

$sel:tags:CreateManagedEndpoint', createManagedEndpoint_tags - The tags of the managed endpoint.

$sel:name:CreateManagedEndpoint', createManagedEndpoint_name - The name of the managed endpoint.

$sel:virtualClusterId:CreateManagedEndpoint', createManagedEndpoint_virtualClusterId - The ID of the virtual cluster for which a managed endpoint is created.

$sel:type':CreateManagedEndpoint', createManagedEndpoint_type - The type of the managed endpoint.

$sel:releaseLabel:CreateManagedEndpoint', createManagedEndpoint_releaseLabel - The Amazon EMR release version.

$sel:executionRoleArn:CreateManagedEndpoint', createManagedEndpoint_executionRoleArn - The ARN of the execution role.

$sel:certificateArn:CreateManagedEndpoint', createManagedEndpoint_certificateArn - The certificate ARN of the managed endpoint.

$sel:clientToken:CreateManagedEndpoint', createManagedEndpoint_clientToken - The client idempotency token for this create call.

Request Lenses

createManagedEndpoint_configurationOverrides :: Lens' CreateManagedEndpoint (Maybe ConfigurationOverrides)

The configuration settings that will be used to override existing configurations.

createManagedEndpoint_tags :: Lens' CreateManagedEndpoint (Maybe (HashMap Text Text))

The tags of the managed endpoint.

createManagedEndpoint_name :: Lens' CreateManagedEndpoint Text

The name of the managed endpoint.

createManagedEndpoint_virtualClusterId :: Lens' CreateManagedEndpoint Text

The ID of the virtual cluster for which a managed endpoint is created.

createManagedEndpoint_type :: Lens' CreateManagedEndpoint Text

The type of the managed endpoint.

createManagedEndpoint_releaseLabel :: Lens' CreateManagedEndpoint Text

The Amazon EMR release version.

createManagedEndpoint_executionRoleArn :: Lens' CreateManagedEndpoint Text

The ARN of the execution role.

createManagedEndpoint_certificateArn :: Lens' CreateManagedEndpoint Text

The certificate ARN of the managed endpoint.

createManagedEndpoint_clientToken :: Lens' CreateManagedEndpoint Text

The client idempotency token for this create call.

Destructuring the Response

data CreateManagedEndpointResponse

See: newCreateManagedEndpointResponse smart constructor.

Constructors

CreateManagedEndpointResponse'

Fields

arn :: Maybe Text

The output contains the ARN of the managed endpoint.

name :: Maybe Text

The output contains the name of the managed endpoint.

id :: Maybe Text

The output contains the ID of the managed endpoint.

virtualClusterId :: Maybe Text

The output contains the ID of the virtual cluster.

httpStatus :: Int

The response's http status code.

newCreateManagedEndpointResponse

Arguments

:: Int
$sel:httpStatus:CreateManagedEndpointResponse'

-> CreateManagedEndpointResponse

Create a value of CreateManagedEndpointResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:arn:CreateManagedEndpointResponse', createManagedEndpointResponse_arn - The output contains the ARN of the managed endpoint.

$sel:name:CreateManagedEndpoint', createManagedEndpointResponse_name - The output contains the name of the managed endpoint.

$sel:id:CreateManagedEndpointResponse', createManagedEndpointResponse_id - The output contains the ID of the managed endpoint.

$sel:virtualClusterId:CreateManagedEndpoint', createManagedEndpointResponse_virtualClusterId - The output contains the ID of the virtual cluster.

$sel:httpStatus:CreateManagedEndpointResponse', createManagedEndpointResponse_httpStatus - The response's http status code.

Response Lenses

createManagedEndpointResponse_arn :: Lens' CreateManagedEndpointResponse (Maybe Text)

The output contains the ARN of the managed endpoint.

createManagedEndpointResponse_name :: Lens' CreateManagedEndpointResponse (Maybe Text)

The output contains the name of the managed endpoint.

createManagedEndpointResponse_id :: Lens' CreateManagedEndpointResponse (Maybe Text)

The output contains the ID of the managed endpoint.

createManagedEndpointResponse_virtualClusterId :: Lens' CreateManagedEndpointResponse (Maybe Text)

The output contains the ID of the virtual cluster.

createManagedEndpointResponse_httpStatus :: Lens' CreateManagedEndpointResponse Int

The response's http status code.