Amazonka.EMRContainers.CreateVirtualCluster

Description

Creates a virtual cluster. Virtual cluster is a managed entity on Amazon EMR on EKS. You can create, describe, list and delete virtual clusters. They do not consume any additional resource in your system. A single virtual cluster maps to a single Kubernetes namespace. Given this relationship, you can model virtual clusters the same way you model Kubernetes namespaces to meet your requirements.

Creating a Request

data CreateVirtualCluster

See: newCreateVirtualCluster smart constructor.

Constructors

CreateVirtualCluster'

Fields

tags :: Maybe (HashMap Text Text)

The tags assigned to the virtual cluster.

name :: Text

The specified name of the virtual cluster.

containerProvider :: ContainerProvider

The container provider of the virtual cluster.

clientToken :: Text

The client token of the virtual cluster.

newCreateVirtualCluster

Arguments

:: Text
$sel:name:CreateVirtualCluster'

-> ContainerProvider
$sel:containerProvider:CreateVirtualCluster'

-> Text
$sel:clientToken:CreateVirtualCluster'

-> CreateVirtualCluster

Create a value of CreateVirtualCluster with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:tags:CreateVirtualCluster', createVirtualCluster_tags - The tags assigned to the virtual cluster.

$sel:name:CreateVirtualCluster', createVirtualCluster_name - The specified name of the virtual cluster.

$sel:containerProvider:CreateVirtualCluster', createVirtualCluster_containerProvider - The container provider of the virtual cluster.

$sel:clientToken:CreateVirtualCluster', createVirtualCluster_clientToken - The client token of the virtual cluster.

Request Lenses

createVirtualCluster_tags :: Lens' CreateVirtualCluster (Maybe (HashMap Text Text))

The tags assigned to the virtual cluster.

createVirtualCluster_name :: Lens' CreateVirtualCluster Text

The specified name of the virtual cluster.

createVirtualCluster_containerProvider :: Lens' CreateVirtualCluster ContainerProvider

The container provider of the virtual cluster.

createVirtualCluster_clientToken :: Lens' CreateVirtualCluster Text

The client token of the virtual cluster.

Destructuring the Response

data CreateVirtualClusterResponse

See: newCreateVirtualClusterResponse smart constructor.

Constructors

CreateVirtualClusterResponse'

Fields

arn :: Maybe Text

This output contains the ARN of virtual cluster.

name :: Maybe Text

This output contains the name of the virtual cluster.

id :: Maybe Text

This output contains the virtual cluster ID.

httpStatus :: Int

The response's http status code.

newCreateVirtualClusterResponse

Arguments

:: Int
$sel:httpStatus:CreateVirtualClusterResponse'

-> CreateVirtualClusterResponse

Create a value of CreateVirtualClusterResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:arn:CreateVirtualClusterResponse', createVirtualClusterResponse_arn - This output contains the ARN of virtual cluster.

$sel:name:CreateVirtualCluster', createVirtualClusterResponse_name - This output contains the name of the virtual cluster.

$sel:id:CreateVirtualClusterResponse', createVirtualClusterResponse_id - This output contains the virtual cluster ID.

$sel:httpStatus:CreateVirtualClusterResponse', createVirtualClusterResponse_httpStatus - The response's http status code.

Response Lenses

createVirtualClusterResponse_arn :: Lens' CreateVirtualClusterResponse (Maybe Text)

This output contains the ARN of virtual cluster.

createVirtualClusterResponse_name :: Lens' CreateVirtualClusterResponse (Maybe Text)

This output contains the name of the virtual cluster.

createVirtualClusterResponse_id :: Lens' CreateVirtualClusterResponse (Maybe Text)

This output contains the virtual cluster ID.

createVirtualClusterResponse_httpStatus :: Lens' CreateVirtualClusterResponse Int

The response's http status code.