Amazonka.EMRContainers.DescribeJobRun

Description

Displays detailed information about a job run. A job run is a unit of work, such as a Spark jar, PySpark script, or SparkSQL query, that you submit to Amazon EMR on EKS.

Creating a Request

data DescribeJobRun

See: newDescribeJobRun smart constructor.

Constructors

DescribeJobRun'

Fields

id :: Text

The ID of the job run request.

virtualClusterId :: Text

The ID of the virtual cluster for which the job run is submitted.

newDescribeJobRun

Arguments

:: Text
$sel:id:DescribeJobRun'

-> Text
$sel:virtualClusterId:DescribeJobRun'

-> DescribeJobRun

Create a value of DescribeJobRun with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:id:DescribeJobRun', describeJobRun_id - The ID of the job run request.

$sel:virtualClusterId:DescribeJobRun', describeJobRun_virtualClusterId - The ID of the virtual cluster for which the job run is submitted.

Request Lenses

describeJobRun_id :: Lens' DescribeJobRun Text

The ID of the job run request.

describeJobRun_virtualClusterId :: Lens' DescribeJobRun Text

The ID of the virtual cluster for which the job run is submitted.

Destructuring the Response

data DescribeJobRunResponse

See: newDescribeJobRunResponse smart constructor.

Constructors

DescribeJobRunResponse'

Fields

jobRun :: Maybe JobRun

The output displays information about a job run.

httpStatus :: Int

The response's http status code.

newDescribeJobRunResponse

Arguments

:: Int
$sel:httpStatus:DescribeJobRunResponse'

-> DescribeJobRunResponse

Create a value of DescribeJobRunResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:jobRun:DescribeJobRunResponse', describeJobRunResponse_jobRun - The output displays information about a job run.

$sel:httpStatus:DescribeJobRunResponse', describeJobRunResponse_httpStatus - The response's http status code.

Response Lenses

describeJobRunResponse_jobRun :: Lens' DescribeJobRunResponse (Maybe JobRun)

The output displays information about a job run.

describeJobRunResponse_httpStatus :: Lens' DescribeJobRunResponse Int

The response's http status code.