Amazonka.EMRContainers.ListManagedEndpoints

Description

Lists managed endpoints based on a set of parameters. A managed endpoint is a gateway that connects EMR Studio to Amazon EMR on EKS so that EMR Studio can communicate with your virtual cluster.

This operation returns paginated results.

Creating a Request

data ListManagedEndpoints

See: newListManagedEndpoints smart constructor.

Constructors

ListManagedEndpoints'

Fields

states :: Maybe [EndpointState]

The states of the managed endpoints.

createdAfter :: Maybe POSIX

The date and time after which the endpoints are created.

types :: Maybe [Text]

The types of the managed endpoints.

nextToken :: Maybe Text

The token for the next set of managed endpoints to return.

maxResults :: Maybe Int

The maximum number of managed endpoints that can be listed.

createdBefore :: Maybe POSIX

The date and time before which the endpoints are created.

virtualClusterId :: Text

The ID of the virtual cluster.

newListManagedEndpoints

Arguments

:: Text
$sel:virtualClusterId:ListManagedEndpoints'

-> ListManagedEndpoints

Create a value of ListManagedEndpoints with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:states:ListManagedEndpoints', listManagedEndpoints_states - The states of the managed endpoints.

$sel:createdAfter:ListManagedEndpoints', listManagedEndpoints_createdAfter - The date and time after which the endpoints are created.

$sel:types:ListManagedEndpoints', listManagedEndpoints_types - The types of the managed endpoints.

$sel:nextToken:ListManagedEndpoints', listManagedEndpoints_nextToken - The token for the next set of managed endpoints to return.

$sel:maxResults:ListManagedEndpoints', listManagedEndpoints_maxResults - The maximum number of managed endpoints that can be listed.

$sel:createdBefore:ListManagedEndpoints', listManagedEndpoints_createdBefore - The date and time before which the endpoints are created.

$sel:virtualClusterId:ListManagedEndpoints', listManagedEndpoints_virtualClusterId - The ID of the virtual cluster.

Request Lenses

listManagedEndpoints_states :: Lens' ListManagedEndpoints (Maybe [EndpointState])

The states of the managed endpoints.

listManagedEndpoints_createdAfter :: Lens' ListManagedEndpoints (Maybe UTCTime)

The date and time after which the endpoints are created.

listManagedEndpoints_types :: Lens' ListManagedEndpoints (Maybe [Text])

The types of the managed endpoints.

listManagedEndpoints_nextToken :: Lens' ListManagedEndpoints (Maybe Text)

The token for the next set of managed endpoints to return.

listManagedEndpoints_maxResults :: Lens' ListManagedEndpoints (Maybe Int)

The maximum number of managed endpoints that can be listed.

listManagedEndpoints_createdBefore :: Lens' ListManagedEndpoints (Maybe UTCTime)

The date and time before which the endpoints are created.

listManagedEndpoints_virtualClusterId :: Lens' ListManagedEndpoints Text

The ID of the virtual cluster.

Destructuring the Response

data ListManagedEndpointsResponse

See: newListManagedEndpointsResponse smart constructor.

Constructors

ListManagedEndpointsResponse'

Fields

nextToken :: Maybe Text

The token for the next set of endpoints to return.

endpoints :: Maybe [Endpoint]

The managed endpoints to be listed.

httpStatus :: Int

The response's http status code.

newListManagedEndpointsResponse

Arguments

:: Int
$sel:httpStatus:ListManagedEndpointsResponse'

-> ListManagedEndpointsResponse

Create a value of ListManagedEndpointsResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:nextToken:ListManagedEndpoints', listManagedEndpointsResponse_nextToken - The token for the next set of endpoints to return.

$sel:endpoints:ListManagedEndpointsResponse', listManagedEndpointsResponse_endpoints - The managed endpoints to be listed.

$sel:httpStatus:ListManagedEndpointsResponse', listManagedEndpointsResponse_httpStatus - The response's http status code.

Response Lenses

listManagedEndpointsResponse_nextToken :: Lens' ListManagedEndpointsResponse (Maybe Text)

The token for the next set of endpoints to return.

listManagedEndpointsResponse_endpoints :: Lens' ListManagedEndpointsResponse (Maybe [Endpoint])

The managed endpoints to be listed.

listManagedEndpointsResponse_httpStatus :: Lens' ListManagedEndpointsResponse Int

The response's http status code.