Amazonka.EMRContainers.ListVirtualClusters

Description

Lists information about the specified virtual cluster. Virtual cluster is a managed entity on Amazon EMR on EKS. You can create, describe, list and delete virtual clusters. They do not consume any additional resource in your system. A single virtual cluster maps to a single Kubernetes namespace. Given this relationship, you can model virtual clusters the same way you model Kubernetes namespaces to meet your requirements.

This operation returns paginated results.

Creating a Request

data ListVirtualClusters

See: newListVirtualClusters smart constructor.

Constructors

ListVirtualClusters'

Fields

states :: Maybe [VirtualClusterState]

The states of the requested virtual clusters.

createdAfter :: Maybe POSIX

The date and time after which the virtual clusters are created.

containerProviderType :: Maybe ContainerProviderType

The container provider type of the virtual cluster. EKS is the only supported type as of now.

nextToken :: Maybe Text

The token for the next set of virtual clusters to return.

containerProviderId :: Maybe Text

The container provider ID of the virtual cluster.

maxResults :: Maybe Int

The maximum number of virtual clusters that can be listed.

createdBefore :: Maybe POSIX

The date and time before which the virtual clusters are created.

newListVirtualClusters :: ListVirtualClusters

Create a value of ListVirtualClusters with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:states:ListVirtualClusters', listVirtualClusters_states - The states of the requested virtual clusters.

$sel:createdAfter:ListVirtualClusters', listVirtualClusters_createdAfter - The date and time after which the virtual clusters are created.

$sel:containerProviderType:ListVirtualClusters', listVirtualClusters_containerProviderType - The container provider type of the virtual cluster. EKS is the only supported type as of now.

$sel:nextToken:ListVirtualClusters', listVirtualClusters_nextToken - The token for the next set of virtual clusters to return.

$sel:containerProviderId:ListVirtualClusters', listVirtualClusters_containerProviderId - The container provider ID of the virtual cluster.

$sel:maxResults:ListVirtualClusters', listVirtualClusters_maxResults - The maximum number of virtual clusters that can be listed.

$sel:createdBefore:ListVirtualClusters', listVirtualClusters_createdBefore - The date and time before which the virtual clusters are created.

Request Lenses

listVirtualClusters_states :: Lens' ListVirtualClusters (Maybe [VirtualClusterState])

The states of the requested virtual clusters.

listVirtualClusters_createdAfter :: Lens' ListVirtualClusters (Maybe UTCTime)

The date and time after which the virtual clusters are created.

listVirtualClusters_containerProviderType :: Lens' ListVirtualClusters (Maybe ContainerProviderType)

The container provider type of the virtual cluster. EKS is the only supported type as of now.

listVirtualClusters_nextToken :: Lens' ListVirtualClusters (Maybe Text)

The token for the next set of virtual clusters to return.

listVirtualClusters_containerProviderId :: Lens' ListVirtualClusters (Maybe Text)

The container provider ID of the virtual cluster.

listVirtualClusters_maxResults :: Lens' ListVirtualClusters (Maybe Int)

The maximum number of virtual clusters that can be listed.

listVirtualClusters_createdBefore :: Lens' ListVirtualClusters (Maybe UTCTime)

The date and time before which the virtual clusters are created.

Destructuring the Response

data ListVirtualClustersResponse

See: newListVirtualClustersResponse smart constructor.

Constructors

ListVirtualClustersResponse'

Fields

nextToken :: Maybe Text

This output displays the token for the next set of virtual clusters.

virtualClusters :: Maybe [VirtualCluster]

This output lists the specified virtual clusters.

httpStatus :: Int

The response's http status code.

newListVirtualClustersResponse

Arguments

:: Int
$sel:httpStatus:ListVirtualClustersResponse'

-> ListVirtualClustersResponse

Create a value of ListVirtualClustersResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:nextToken:ListVirtualClusters', listVirtualClustersResponse_nextToken - This output displays the token for the next set of virtual clusters.

$sel:virtualClusters:ListVirtualClustersResponse', listVirtualClustersResponse_virtualClusters - This output lists the specified virtual clusters.

$sel:httpStatus:ListVirtualClustersResponse', listVirtualClustersResponse_httpStatus - The response's http status code.

Response Lenses

listVirtualClustersResponse_nextToken :: Lens' ListVirtualClustersResponse (Maybe Text)

This output displays the token for the next set of virtual clusters.

listVirtualClustersResponse_virtualClusters :: Lens' ListVirtualClustersResponse (Maybe [VirtualCluster])

This output lists the specified virtual clusters.

listVirtualClustersResponse_httpStatus :: Lens' ListVirtualClustersResponse Int

The response's http status code.