Amazonka.EMRContainers.StartJobRun

Description

Starts a job run. A job run is a unit of work, such as a Spark jar, PySpark script, or SparkSQL query, that you submit to Amazon EMR on EKS.

Creating a Request

data StartJobRun

See: newStartJobRun smart constructor.

Constructors

StartJobRun'

Fields

configurationOverrides :: Maybe ConfigurationOverrides

The configuration overrides for the job run.

name :: Maybe Text

The name of the job run.

tags :: Maybe (HashMap Text Text)

The tags assigned to job runs.

virtualClusterId :: Text

The virtual cluster ID for which the job run request is submitted.

clientToken :: Text

The client idempotency token of the job run request.

executionRoleArn :: Text

The execution role ARN for the job run.

releaseLabel :: Text

The Amazon EMR release version to use for the job run.

jobDriver :: JobDriver

The job driver for the job run.

newStartJobRun

Arguments

:: Text
$sel:virtualClusterId:StartJobRun'

-> Text
$sel:clientToken:StartJobRun'

-> Text
$sel:executionRoleArn:StartJobRun'

-> Text
$sel:releaseLabel:StartJobRun'

-> JobDriver
$sel:jobDriver:StartJobRun'

-> StartJobRun

Create a value of StartJobRun with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:configurationOverrides:StartJobRun', startJobRun_configurationOverrides - The configuration overrides for the job run.

$sel:name:StartJobRun', startJobRun_name - The name of the job run.

$sel:tags:StartJobRun', startJobRun_tags - The tags assigned to job runs.

$sel:virtualClusterId:StartJobRun', startJobRun_virtualClusterId - The virtual cluster ID for which the job run request is submitted.

$sel:clientToken:StartJobRun', startJobRun_clientToken - The client idempotency token of the job run request.

$sel:executionRoleArn:StartJobRun', startJobRun_executionRoleArn - The execution role ARN for the job run.

$sel:releaseLabel:StartJobRun', startJobRun_releaseLabel - The Amazon EMR release version to use for the job run.

$sel:jobDriver:StartJobRun', startJobRun_jobDriver - The job driver for the job run.

Request Lenses

startJobRun_configurationOverrides :: Lens' StartJobRun (Maybe ConfigurationOverrides)

The configuration overrides for the job run.

startJobRun_name :: Lens' StartJobRun (Maybe Text)

The name of the job run.

startJobRun_tags :: Lens' StartJobRun (Maybe (HashMap Text Text))

The tags assigned to job runs.

startJobRun_virtualClusterId :: Lens' StartJobRun Text

The virtual cluster ID for which the job run request is submitted.

startJobRun_clientToken :: Lens' StartJobRun Text

The client idempotency token of the job run request.

startJobRun_executionRoleArn :: Lens' StartJobRun Text

The execution role ARN for the job run.

startJobRun_releaseLabel :: Lens' StartJobRun Text

The Amazon EMR release version to use for the job run.

startJobRun_jobDriver :: Lens' StartJobRun JobDriver

The job driver for the job run.

Destructuring the Response

data StartJobRunResponse

See: newStartJobRunResponse smart constructor.

Constructors

StartJobRunResponse'

Fields

arn :: Maybe Text

This output lists the ARN of job run.

name :: Maybe Text

This output displays the name of the started job run.

id :: Maybe Text

This output displays the started job run ID.

virtualClusterId :: Maybe Text

This output displays the virtual cluster ID for which the job run was submitted.

httpStatus :: Int

The response's http status code.

newStartJobRunResponse

Arguments

:: Int
$sel:httpStatus:StartJobRunResponse'

-> StartJobRunResponse

Create a value of StartJobRunResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:arn:StartJobRunResponse', startJobRunResponse_arn - This output lists the ARN of job run.

$sel:name:StartJobRun', startJobRunResponse_name - This output displays the name of the started job run.

$sel:id:StartJobRunResponse', startJobRunResponse_id - This output displays the started job run ID.

$sel:virtualClusterId:StartJobRun', startJobRunResponse_virtualClusterId - This output displays the virtual cluster ID for which the job run was submitted.

$sel:httpStatus:StartJobRunResponse', startJobRunResponse_httpStatus - The response's http status code.

Response Lenses

startJobRunResponse_arn :: Lens' StartJobRunResponse (Maybe Text)

This output lists the ARN of job run.

startJobRunResponse_name :: Lens' StartJobRunResponse (Maybe Text)

This output displays the name of the started job run.

startJobRunResponse_id :: Lens' StartJobRunResponse (Maybe Text)

This output displays the started job run ID.

startJobRunResponse_virtualClusterId :: Lens' StartJobRunResponse (Maybe Text)

This output displays the virtual cluster ID for which the job run was submitted.

startJobRunResponse_httpStatus :: Lens' StartJobRunResponse Int

The response's http status code.