Amazonka.EMRContainers.Types

Description

Service Configuration

defaultService :: Service

API version 2020-10-01 of the Amazon EMR Containers SDK configuration.

Errors

_ValidationException :: AsError a => Getting (First ServiceError) a ServiceError

There are invalid parameters in the client request.

_InternalServerException :: AsError a => Getting (First ServiceError) a ServiceError

This is an internal server exception.

_ResourceNotFoundException :: AsError a => Getting (First ServiceError) a ServiceError

The specified resource was not found.

ContainerProviderType

newtype ContainerProviderType

Constructors

ContainerProviderType'

Fields

fromContainerProviderType :: Text

Bundled Patterns

pattern ContainerProviderType_EKS :: ContainerProviderType

EndpointState

newtype EndpointState

Constructors

EndpointState'

Fields

fromEndpointState :: Text

Bundled Patterns

pattern EndpointState_ACTIVE :: EndpointState

pattern EndpointState_CREATING :: EndpointState

pattern EndpointState_TERMINATED :: EndpointState

pattern EndpointState_TERMINATED_WITH_ERRORS :: EndpointState

pattern EndpointState_TERMINATING :: EndpointState

FailureReason

newtype FailureReason

Constructors

FailureReason'

Fields

fromFailureReason :: Text

Bundled Patterns

pattern FailureReason_CLUSTER_UNAVAILABLE :: FailureReason

pattern FailureReason_INTERNAL_ERROR :: FailureReason

pattern FailureReason_USER_ERROR :: FailureReason

pattern FailureReason_VALIDATION_ERROR :: FailureReason

JobRunState

newtype JobRunState

Constructors

JobRunState'

Fields

fromJobRunState :: Text

Bundled Patterns

pattern JobRunState_CANCELLED :: JobRunState

pattern JobRunState_CANCEL_PENDING :: JobRunState

pattern JobRunState_COMPLETED :: JobRunState

pattern JobRunState_FAILED :: JobRunState

pattern JobRunState_PENDING :: JobRunState

pattern JobRunState_RUNNING :: JobRunState

pattern JobRunState_SUBMITTED :: JobRunState

PersistentAppUI

newtype PersistentAppUI

Constructors

PersistentAppUI'

Fields

fromPersistentAppUI :: Text

Bundled Patterns

pattern PersistentAppUI_DISABLED :: PersistentAppUI

pattern PersistentAppUI_ENABLED :: PersistentAppUI

VirtualClusterState

newtype VirtualClusterState

Constructors

VirtualClusterState'

Fields

fromVirtualClusterState :: Text

Bundled Patterns

pattern VirtualClusterState_ARRESTED :: VirtualClusterState

pattern VirtualClusterState_RUNNING :: VirtualClusterState

pattern VirtualClusterState_TERMINATED :: VirtualClusterState

pattern VirtualClusterState_TERMINATING :: VirtualClusterState

CloudWatchMonitoringConfiguration

data CloudWatchMonitoringConfiguration

A configuration for CloudWatch monitoring. You can configure your jobs to send log information to CloudWatch Logs.

See: newCloudWatchMonitoringConfiguration smart constructor.

Constructors

CloudWatchMonitoringConfiguration'

Fields

logStreamNamePrefix :: Maybe Text

The specified name prefix for log streams.

logGroupName :: Text

The name of the log group for log publishing.

newCloudWatchMonitoringConfiguration

Arguments

:: Text
$sel:logGroupName:CloudWatchMonitoringConfiguration'

-> CloudWatchMonitoringConfiguration

Create a value of CloudWatchMonitoringConfiguration with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:logStreamNamePrefix:CloudWatchMonitoringConfiguration', cloudWatchMonitoringConfiguration_logStreamNamePrefix - The specified name prefix for log streams.

$sel:logGroupName:CloudWatchMonitoringConfiguration', cloudWatchMonitoringConfiguration_logGroupName - The name of the log group for log publishing.

cloudWatchMonitoringConfiguration_logStreamNamePrefix :: Lens' CloudWatchMonitoringConfiguration (Maybe Text)

The specified name prefix for log streams.

cloudWatchMonitoringConfiguration_logGroupName :: Lens' CloudWatchMonitoringConfiguration Text

The name of the log group for log publishing.

Configuration

data Configuration

A configuration specification to be used when provisioning virtual clusters, which can include configurations for applications and software bundled with Amazon EMR on EKS. A configuration consists of a classification, properties, and optional nested configurations. A classification refers to an application-specific configuration file. Properties are the settings you want to change in that file.

See: newConfiguration smart constructor.

Constructors

Configuration'

Fields

configurations :: Maybe [Configuration]

A list of additional configurations to apply within a configuration object.

properties :: Maybe (Sensitive (HashMap Text Text))

A set of properties specified within a configuration classification.

classification :: Text

The classification within a configuration.

newConfiguration

Arguments

:: Text
$sel:classification:Configuration'

-> Configuration

Create a value of Configuration with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:configurations:Configuration', configuration_configurations - A list of additional configurations to apply within a configuration object.

$sel:properties:Configuration', configuration_properties - A set of properties specified within a configuration classification.

$sel:classification:Configuration', configuration_classification - The classification within a configuration.

configuration_configurations :: Lens' Configuration (Maybe [Configuration])

A list of additional configurations to apply within a configuration object.

configuration_properties :: Lens' Configuration (Maybe (HashMap Text Text))

A set of properties specified within a configuration classification.

configuration_classification :: Lens' Configuration Text

The classification within a configuration.

ConfigurationOverrides

data ConfigurationOverrides

A configuration specification to be used to override existing configurations.

See: newConfigurationOverrides smart constructor.

Constructors

ConfigurationOverrides'

Fields

monitoringConfiguration :: Maybe MonitoringConfiguration

The configurations for monitoring.

applicationConfiguration :: Maybe [Configuration]

The configurations for the application running by the job run.

newConfigurationOverrides :: ConfigurationOverrides

Create a value of ConfigurationOverrides with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:monitoringConfiguration:ConfigurationOverrides', configurationOverrides_monitoringConfiguration - The configurations for monitoring.

$sel:applicationConfiguration:ConfigurationOverrides', configurationOverrides_applicationConfiguration - The configurations for the application running by the job run.

configurationOverrides_monitoringConfiguration :: Lens' ConfigurationOverrides (Maybe MonitoringConfiguration)

The configurations for monitoring.

configurationOverrides_applicationConfiguration :: Lens' ConfigurationOverrides (Maybe [Configuration])

The configurations for the application running by the job run.

ContainerInfo

data ContainerInfo

The information about the container used for a job run or a managed endpoint.

See: newContainerInfo smart constructor.

Constructors

ContainerInfo'

Fields

eksInfo :: Maybe EksInfo

The information about the EKS cluster.

newContainerInfo :: ContainerInfo

Create a value of ContainerInfo with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:eksInfo:ContainerInfo', containerInfo_eksInfo - The information about the EKS cluster.

containerInfo_eksInfo :: Lens' ContainerInfo (Maybe EksInfo)

The information about the EKS cluster.

ContainerProvider

data ContainerProvider

The information about the container provider.

See: newContainerProvider smart constructor.

Constructors

ContainerProvider'

Fields

info :: Maybe ContainerInfo

The information about the container cluster.

type' :: ContainerProviderType

The type of the container provider. EKS is the only supported type as of now.

id :: Text

The ID of the container cluster.

newContainerProvider

Arguments

:: ContainerProviderType
$sel:type':ContainerProvider'

-> Text
$sel:id:ContainerProvider'

-> ContainerProvider

Create a value of ContainerProvider with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:info:ContainerProvider', containerProvider_info - The information about the container cluster.

$sel:type':ContainerProvider', containerProvider_type - The type of the container provider. EKS is the only supported type as of now.

$sel:id:ContainerProvider', containerProvider_id - The ID of the container cluster.

containerProvider_info :: Lens' ContainerProvider (Maybe ContainerInfo)

The information about the container cluster.

containerProvider_type :: Lens' ContainerProvider ContainerProviderType

The type of the container provider. EKS is the only supported type as of now.

containerProvider_id :: Lens' ContainerProvider Text

The ID of the container cluster.

EksInfo

data EksInfo

The information about the EKS cluster.

See: newEksInfo smart constructor.

Constructors

EksInfo'

Fields

namespace :: Maybe Text

The namespaces of the EKS cluster.

newEksInfo :: EksInfo

Create a value of EksInfo with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:namespace:EksInfo', eksInfo_namespace - The namespaces of the EKS cluster.

eksInfo_namespace :: Lens' EksInfo (Maybe Text)

The namespaces of the EKS cluster.

Endpoint

data Endpoint

This entity represents the endpoint that is managed by Amazon EMR on EKS.

See: newEndpoint smart constructor.

Constructors

Endpoint'

Fields

failureReason :: Maybe FailureReason

The reasons why the endpoint has failed.

state :: Maybe EndpointState

The state of the endpoint.

arn :: Maybe Text

The ARN of the endpoint.

createdAt :: Maybe POSIX

The date and time when the endpoint was created.

subnetIds :: Maybe [Text]

The subnet IDs of the endpoint.

stateDetails :: Maybe Text

Additional details of the endpoint state.

certificateArn :: Maybe Text

The certificate ARN of the endpoint.

executionRoleArn :: Maybe Text

The execution role ARN of the endpoint.

securityGroup :: Maybe Text

The security group configuration of the endpoint.

configurationOverrides :: Maybe ConfigurationOverrides

The configuration settings that are used to override existing configurations for endpoints.

name :: Maybe Text

The name of the endpoint.

releaseLabel :: Maybe Text

The EMR release version to be used for the endpoint.

id :: Maybe Text

The ID of the endpoint.

type' :: Maybe Text

The type of the endpoint.

serverUrl :: Maybe Text

The server URL of the endpoint.

virtualClusterId :: Maybe Text

The ID of the endpoint's virtual cluster.

tags :: Maybe (HashMap Text Text)

The tags of the endpoint.

newEndpoint :: Endpoint

Create a value of Endpoint with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:failureReason:Endpoint', endpoint_failureReason - The reasons why the endpoint has failed.

$sel:state:Endpoint', endpoint_state - The state of the endpoint.

$sel:arn:Endpoint', endpoint_arn - The ARN of the endpoint.

$sel:createdAt:Endpoint', endpoint_createdAt - The date and time when the endpoint was created.

$sel:subnetIds:Endpoint', endpoint_subnetIds - The subnet IDs of the endpoint.

$sel:stateDetails:Endpoint', endpoint_stateDetails - Additional details of the endpoint state.

$sel:certificateArn:Endpoint', endpoint_certificateArn - The certificate ARN of the endpoint.

$sel:executionRoleArn:Endpoint', endpoint_executionRoleArn - The execution role ARN of the endpoint.

$sel:securityGroup:Endpoint', endpoint_securityGroup - The security group configuration of the endpoint.

$sel:configurationOverrides:Endpoint', endpoint_configurationOverrides - The configuration settings that are used to override existing configurations for endpoints.

$sel:name:Endpoint', endpoint_name - The name of the endpoint.

$sel:releaseLabel:Endpoint', endpoint_releaseLabel - The EMR release version to be used for the endpoint.

$sel:id:Endpoint', endpoint_id - The ID of the endpoint.

$sel:type':Endpoint', endpoint_type - The type of the endpoint.

$sel:serverUrl:Endpoint', endpoint_serverUrl - The server URL of the endpoint.

$sel:virtualClusterId:Endpoint', endpoint_virtualClusterId - The ID of the endpoint's virtual cluster.

$sel:tags:Endpoint', endpoint_tags - The tags of the endpoint.

endpoint_failureReason :: Lens' Endpoint (Maybe FailureReason)

The reasons why the endpoint has failed.

endpoint_state :: Lens' Endpoint (Maybe EndpointState)

The state of the endpoint.

endpoint_arn :: Lens' Endpoint (Maybe Text)

The ARN of the endpoint.

endpoint_createdAt :: Lens' Endpoint (Maybe UTCTime)

The date and time when the endpoint was created.

endpoint_subnetIds :: Lens' Endpoint (Maybe [Text])

The subnet IDs of the endpoint.

endpoint_stateDetails :: Lens' Endpoint (Maybe Text)

Additional details of the endpoint state.

endpoint_certificateArn :: Lens' Endpoint (Maybe Text)

The certificate ARN of the endpoint.

endpoint_executionRoleArn :: Lens' Endpoint (Maybe Text)

The execution role ARN of the endpoint.

endpoint_securityGroup :: Lens' Endpoint (Maybe Text)

The security group configuration of the endpoint.

endpoint_configurationOverrides :: Lens' Endpoint (Maybe ConfigurationOverrides)

The configuration settings that are used to override existing configurations for endpoints.

endpoint_name :: Lens' Endpoint (Maybe Text)

The name of the endpoint.

endpoint_releaseLabel :: Lens' Endpoint (Maybe Text)

The EMR release version to be used for the endpoint.

endpoint_id :: Lens' Endpoint (Maybe Text)

The ID of the endpoint.

endpoint_type :: Lens' Endpoint (Maybe Text)

The type of the endpoint.

endpoint_serverUrl :: Lens' Endpoint (Maybe Text)

The server URL of the endpoint.

endpoint_virtualClusterId :: Lens' Endpoint (Maybe Text)

The ID of the endpoint's virtual cluster.

endpoint_tags :: Lens' Endpoint (Maybe (HashMap Text Text))

The tags of the endpoint.

JobDriver

data JobDriver

Specify the driver that the job runs on.

See: newJobDriver smart constructor.

Constructors

JobDriver'

Fields

sparkSubmitJobDriver :: Maybe SparkSubmitJobDriver

The job driver parameters specified for spark submit.

newJobDriver :: JobDriver

Create a value of JobDriver with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:sparkSubmitJobDriver:JobDriver', jobDriver_sparkSubmitJobDriver - The job driver parameters specified for spark submit.

jobDriver_sparkSubmitJobDriver :: Lens' JobDriver (Maybe SparkSubmitJobDriver)

The job driver parameters specified for spark submit.

JobRun

data JobRun

This entity describes a job run. A job run is a unit of work, such as a Spark jar, PySpark script, or SparkSQL query, that you submit to Amazon EMR on EKS.

See: newJobRun smart constructor.

Constructors

JobRun'

Fields

failureReason :: Maybe FailureReason

The reasons why the job run has failed.

state :: Maybe JobRunState

The state of the job run.

clientToken :: Maybe Text

The client token used to start a job run.

arn :: Maybe Text

The ARN of job run.

createdAt :: Maybe POSIX

The date and time when the job run was created.

stateDetails :: Maybe Text

Additional details of the job run state.

createdBy :: Maybe Text

The user who created the job run.

executionRoleArn :: Maybe Text

The execution role ARN of the job run.

jobDriver :: Maybe JobDriver

Parameters of job driver for the job run.

configurationOverrides :: Maybe ConfigurationOverrides

The configuration settings that are used to override default configuration.

finishedAt :: Maybe POSIX

The date and time when the job run has finished.

name :: Maybe Text

The name of the job run.

releaseLabel :: Maybe Text

The release version of Amazon EMR.

id :: Maybe Text

The ID of the job run.

virtualClusterId :: Maybe Text

The ID of the job run's virtual cluster.

tags :: Maybe (HashMap Text Text)

The assigned tags of the job run.

newJobRun :: JobRun

Create a value of JobRun with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:failureReason:JobRun', jobRun_failureReason - The reasons why the job run has failed.

$sel:state:JobRun', jobRun_state - The state of the job run.

$sel:clientToken:JobRun', jobRun_clientToken - The client token used to start a job run.

$sel:arn:JobRun', jobRun_arn - The ARN of job run.

$sel:createdAt:JobRun', jobRun_createdAt - The date and time when the job run was created.

$sel:stateDetails:JobRun', jobRun_stateDetails - Additional details of the job run state.

$sel:createdBy:JobRun', jobRun_createdBy - The user who created the job run.

$sel:executionRoleArn:JobRun', jobRun_executionRoleArn - The execution role ARN of the job run.

$sel:jobDriver:JobRun', jobRun_jobDriver - Parameters of job driver for the job run.

$sel:configurationOverrides:JobRun', jobRun_configurationOverrides - The configuration settings that are used to override default configuration.

$sel:finishedAt:JobRun', jobRun_finishedAt - The date and time when the job run has finished.

$sel:name:JobRun', jobRun_name - The name of the job run.

$sel:releaseLabel:JobRun', jobRun_releaseLabel - The release version of Amazon EMR.

$sel:id:JobRun', jobRun_id - The ID of the job run.

$sel:virtualClusterId:JobRun', jobRun_virtualClusterId - The ID of the job run's virtual cluster.

$sel:tags:JobRun', jobRun_tags - The assigned tags of the job run.

jobRun_failureReason :: Lens' JobRun (Maybe FailureReason)

The reasons why the job run has failed.

jobRun_state :: Lens' JobRun (Maybe JobRunState)

The state of the job run.

jobRun_clientToken :: Lens' JobRun (Maybe Text)

The client token used to start a job run.

jobRun_arn :: Lens' JobRun (Maybe Text)

The ARN of job run.

jobRun_createdAt :: Lens' JobRun (Maybe UTCTime)

The date and time when the job run was created.

jobRun_stateDetails :: Lens' JobRun (Maybe Text)

Additional details of the job run state.

jobRun_createdBy :: Lens' JobRun (Maybe Text)

The user who created the job run.

jobRun_executionRoleArn :: Lens' JobRun (Maybe Text)

The execution role ARN of the job run.

jobRun_jobDriver :: Lens' JobRun (Maybe JobDriver)

Parameters of job driver for the job run.

jobRun_configurationOverrides :: Lens' JobRun (Maybe ConfigurationOverrides)

The configuration settings that are used to override default configuration.

jobRun_finishedAt :: Lens' JobRun (Maybe UTCTime)

The date and time when the job run has finished.

jobRun_name :: Lens' JobRun (Maybe Text)

The name of the job run.

jobRun_releaseLabel :: Lens' JobRun (Maybe Text)

The release version of Amazon EMR.

jobRun_id :: Lens' JobRun (Maybe Text)

The ID of the job run.

jobRun_virtualClusterId :: Lens' JobRun (Maybe Text)

The ID of the job run's virtual cluster.

jobRun_tags :: Lens' JobRun (Maybe (HashMap Text Text))

The assigned tags of the job run.

MonitoringConfiguration

data MonitoringConfiguration

Configuration setting for monitoring.

See: newMonitoringConfiguration smart constructor.

Constructors

MonitoringConfiguration'

Fields

persistentAppUI :: Maybe PersistentAppUI

Monitoring configurations for the persistent application UI.

s3MonitoringConfiguration :: Maybe S3MonitoringConfiguration

Amazon S3 configuration for monitoring log publishing.

cloudWatchMonitoringConfiguration :: Maybe CloudWatchMonitoringConfiguration

Monitoring configurations for CloudWatch.

newMonitoringConfiguration :: MonitoringConfiguration

Create a value of MonitoringConfiguration with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:persistentAppUI:MonitoringConfiguration', monitoringConfiguration_persistentAppUI - Monitoring configurations for the persistent application UI.

$sel:s3MonitoringConfiguration:MonitoringConfiguration', monitoringConfiguration_s3MonitoringConfiguration - Amazon S3 configuration for monitoring log publishing.

$sel:cloudWatchMonitoringConfiguration:MonitoringConfiguration', monitoringConfiguration_cloudWatchMonitoringConfiguration - Monitoring configurations for CloudWatch.

monitoringConfiguration_persistentAppUI :: Lens' MonitoringConfiguration (Maybe PersistentAppUI)

Monitoring configurations for the persistent application UI.

monitoringConfiguration_s3MonitoringConfiguration :: Lens' MonitoringConfiguration (Maybe S3MonitoringConfiguration)

Amazon S3 configuration for monitoring log publishing.

monitoringConfiguration_cloudWatchMonitoringConfiguration :: Lens' MonitoringConfiguration (Maybe CloudWatchMonitoringConfiguration)

Monitoring configurations for CloudWatch.

S3MonitoringConfiguration

data S3MonitoringConfiguration

Amazon S3 configuration for monitoring log publishing. You can configure your jobs to send log information to Amazon S3.

See: newS3MonitoringConfiguration smart constructor.

Constructors

S3MonitoringConfiguration'

Fields

logUri :: Text

Amazon S3 destination URI for log publishing.

newS3MonitoringConfiguration

Arguments

:: Text
$sel:logUri:S3MonitoringConfiguration'

-> S3MonitoringConfiguration

Create a value of S3MonitoringConfiguration with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:logUri:S3MonitoringConfiguration', s3MonitoringConfiguration_logUri - Amazon S3 destination URI for log publishing.

s3MonitoringConfiguration_logUri :: Lens' S3MonitoringConfiguration Text

Amazon S3 destination URI for log publishing.

SparkSubmitJobDriver

data SparkSubmitJobDriver

The information about job driver for Spark submit.

See: newSparkSubmitJobDriver smart constructor.

Constructors

SparkSubmitJobDriver'

Fields

sparkSubmitParameters :: Maybe (Sensitive Text)

The Spark submit parameters that are used for job runs.

entryPointArguments :: Maybe [Sensitive Text]

The arguments for job application.

entryPoint :: Sensitive Text

The entry point of job application.

newSparkSubmitJobDriver

Arguments

:: Text
$sel:entryPoint:SparkSubmitJobDriver'

-> SparkSubmitJobDriver

Create a value of SparkSubmitJobDriver with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:sparkSubmitParameters:SparkSubmitJobDriver', sparkSubmitJobDriver_sparkSubmitParameters - The Spark submit parameters that are used for job runs.

$sel:entryPointArguments:SparkSubmitJobDriver', sparkSubmitJobDriver_entryPointArguments - The arguments for job application.

$sel:entryPoint:SparkSubmitJobDriver', sparkSubmitJobDriver_entryPoint - The entry point of job application.

sparkSubmitJobDriver_sparkSubmitParameters :: Lens' SparkSubmitJobDriver (Maybe Text)

The Spark submit parameters that are used for job runs.

sparkSubmitJobDriver_entryPointArguments :: Lens' SparkSubmitJobDriver (Maybe [Text])

The arguments for job application.

sparkSubmitJobDriver_entryPoint :: Lens' SparkSubmitJobDriver Text

The entry point of job application.

VirtualCluster

data VirtualCluster

This entity describes a virtual cluster. A virtual cluster is a Kubernetes namespace that Amazon EMR is registered with. Amazon EMR uses virtual clusters to run jobs and host endpoints. Multiple virtual clusters can be backed by the same physical cluster. However, each virtual cluster maps to one namespace on an EKS cluster. Virtual clusters do not create any active resources that contribute to your bill or that require lifecycle management outside the service.

See: newVirtualCluster smart constructor.

Constructors

VirtualCluster'

Fields

state :: Maybe VirtualClusterState

The state of the virtual cluster.

arn :: Maybe Text

The ARN of the virtual cluster.

createdAt :: Maybe POSIX

The date and time when the virtual cluster is created.

name :: Maybe Text

The name of the virtual cluster.

id :: Maybe Text

The ID of the virtual cluster.

containerProvider :: Maybe ContainerProvider

The container provider of the virtual cluster.

tags :: Maybe (HashMap Text Text)

The assigned tags of the virtual cluster.

newVirtualCluster :: VirtualCluster

Create a value of VirtualCluster with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:state:VirtualCluster', virtualCluster_state - The state of the virtual cluster.

$sel:arn:VirtualCluster', virtualCluster_arn - The ARN of the virtual cluster.

$sel:createdAt:VirtualCluster', virtualCluster_createdAt - The date and time when the virtual cluster is created.

$sel:name:VirtualCluster', virtualCluster_name - The name of the virtual cluster.

$sel:id:VirtualCluster', virtualCluster_id - The ID of the virtual cluster.

$sel:containerProvider:VirtualCluster', virtualCluster_containerProvider - The container provider of the virtual cluster.

$sel:tags:VirtualCluster', virtualCluster_tags - The assigned tags of the virtual cluster.

virtualCluster_state :: Lens' VirtualCluster (Maybe VirtualClusterState)

The state of the virtual cluster.

virtualCluster_arn :: Lens' VirtualCluster (Maybe Text)

The ARN of the virtual cluster.

virtualCluster_createdAt :: Lens' VirtualCluster (Maybe UTCTime)

The date and time when the virtual cluster is created.

virtualCluster_name :: Lens' VirtualCluster (Maybe Text)

The name of the virtual cluster.

virtualCluster_id :: Lens' VirtualCluster (Maybe Text)

The ID of the virtual cluster.

virtualCluster_containerProvider :: Lens' VirtualCluster (Maybe ContainerProvider)

The container provider of the virtual cluster.

virtualCluster_tags :: Lens' VirtualCluster (Maybe (HashMap Text Text))

The assigned tags of the virtual cluster.