Amazonka.EMRContainers.Types.ContainerProvider

Description

Documentation

data ContainerProvider

The information about the container provider.

See: newContainerProvider smart constructor.

Constructors

ContainerProvider'

Fields

info :: Maybe ContainerInfo

The information about the container cluster.

type' :: ContainerProviderType

The type of the container provider. EKS is the only supported type as of now.

id :: Text

The ID of the container cluster.

newContainerProvider

Arguments

:: ContainerProviderType
$sel:type':ContainerProvider'

-> Text
$sel:id:ContainerProvider'

-> ContainerProvider

Create a value of ContainerProvider with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:info:ContainerProvider', containerProvider_info - The information about the container cluster.

$sel:type':ContainerProvider', containerProvider_type - The type of the container provider. EKS is the only supported type as of now.

$sel:id:ContainerProvider', containerProvider_id - The ID of the container cluster.

containerProvider_info :: Lens' ContainerProvider (Maybe ContainerInfo)

The information about the container cluster.

containerProvider_type :: Lens' ContainerProvider ContainerProviderType

The type of the container provider. EKS is the only supported type as of now.

containerProvider_id :: Lens' ContainerProvider Text

The ID of the container cluster.