Amazonka.EMRContainers.Types.JobRun

Description

Documentation

data JobRun

This entity describes a job run. A job run is a unit of work, such as a Spark jar, PySpark script, or SparkSQL query, that you submit to Amazon EMR on EKS.

See: newJobRun smart constructor.

Constructors

JobRun'

Fields

failureReason :: Maybe FailureReason

The reasons why the job run has failed.

state :: Maybe JobRunState

The state of the job run.

clientToken :: Maybe Text

The client token used to start a job run.

arn :: Maybe Text

The ARN of job run.

createdAt :: Maybe POSIX

The date and time when the job run was created.

stateDetails :: Maybe Text

Additional details of the job run state.

createdBy :: Maybe Text

The user who created the job run.

executionRoleArn :: Maybe Text

The execution role ARN of the job run.

jobDriver :: Maybe JobDriver

Parameters of job driver for the job run.

configurationOverrides :: Maybe ConfigurationOverrides

The configuration settings that are used to override default configuration.

finishedAt :: Maybe POSIX

The date and time when the job run has finished.

name :: Maybe Text

The name of the job run.

releaseLabel :: Maybe Text

The release version of Amazon EMR.

id :: Maybe Text

The ID of the job run.

virtualClusterId :: Maybe Text

The ID of the job run's virtual cluster.

tags :: Maybe (HashMap Text Text)

The assigned tags of the job run.

newJobRun :: JobRun

Create a value of JobRun with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:failureReason:JobRun', jobRun_failureReason - The reasons why the job run has failed.

$sel:state:JobRun', jobRun_state - The state of the job run.

$sel:clientToken:JobRun', jobRun_clientToken - The client token used to start a job run.

$sel:arn:JobRun', jobRun_arn - The ARN of job run.

$sel:createdAt:JobRun', jobRun_createdAt - The date and time when the job run was created.

$sel:stateDetails:JobRun', jobRun_stateDetails - Additional details of the job run state.

$sel:createdBy:JobRun', jobRun_createdBy - The user who created the job run.

$sel:executionRoleArn:JobRun', jobRun_executionRoleArn - The execution role ARN of the job run.

$sel:jobDriver:JobRun', jobRun_jobDriver - Parameters of job driver for the job run.

$sel:configurationOverrides:JobRun', jobRun_configurationOverrides - The configuration settings that are used to override default configuration.

$sel:finishedAt:JobRun', jobRun_finishedAt - The date and time when the job run has finished.

$sel:name:JobRun', jobRun_name - The name of the job run.

$sel:releaseLabel:JobRun', jobRun_releaseLabel - The release version of Amazon EMR.

$sel:id:JobRun', jobRun_id - The ID of the job run.

$sel:virtualClusterId:JobRun', jobRun_virtualClusterId - The ID of the job run's virtual cluster.

$sel:tags:JobRun', jobRun_tags - The assigned tags of the job run.

jobRun_failureReason :: Lens' JobRun (Maybe FailureReason)

The reasons why the job run has failed.

jobRun_state :: Lens' JobRun (Maybe JobRunState)

The state of the job run.

jobRun_clientToken :: Lens' JobRun (Maybe Text)

The client token used to start a job run.

jobRun_arn :: Lens' JobRun (Maybe Text)

The ARN of job run.

jobRun_createdAt :: Lens' JobRun (Maybe UTCTime)

The date and time when the job run was created.

jobRun_stateDetails :: Lens' JobRun (Maybe Text)

Additional details of the job run state.

jobRun_createdBy :: Lens' JobRun (Maybe Text)

The user who created the job run.

jobRun_executionRoleArn :: Lens' JobRun (Maybe Text)

The execution role ARN of the job run.

jobRun_jobDriver :: Lens' JobRun (Maybe JobDriver)

Parameters of job driver for the job run.

jobRun_configurationOverrides :: Lens' JobRun (Maybe ConfigurationOverrides)

The configuration settings that are used to override default configuration.

jobRun_finishedAt :: Lens' JobRun (Maybe UTCTime)

The date and time when the job run has finished.

jobRun_name :: Lens' JobRun (Maybe Text)

The name of the job run.

jobRun_releaseLabel :: Lens' JobRun (Maybe Text)

The release version of Amazon EMR.

jobRun_id :: Lens' JobRun (Maybe Text)

The ID of the job run.

jobRun_virtualClusterId :: Lens' JobRun (Maybe Text)

The ID of the job run's virtual cluster.

jobRun_tags :: Lens' JobRun (Maybe (HashMap Text Text))

The assigned tags of the job run.