Amazonka.EMRContainers.Types.VirtualCluster

Description

Documentation

data VirtualCluster

This entity describes a virtual cluster. A virtual cluster is a Kubernetes namespace that Amazon EMR is registered with. Amazon EMR uses virtual clusters to run jobs and host endpoints. Multiple virtual clusters can be backed by the same physical cluster. However, each virtual cluster maps to one namespace on an EKS cluster. Virtual clusters do not create any active resources that contribute to your bill or that require lifecycle management outside the service.

See: newVirtualCluster smart constructor.

Constructors

VirtualCluster'

Fields

state :: Maybe VirtualClusterState

The state of the virtual cluster.

arn :: Maybe Text

The ARN of the virtual cluster.

createdAt :: Maybe POSIX

The date and time when the virtual cluster is created.

name :: Maybe Text

The name of the virtual cluster.

id :: Maybe Text

The ID of the virtual cluster.

containerProvider :: Maybe ContainerProvider

The container provider of the virtual cluster.

tags :: Maybe (HashMap Text Text)

The assigned tags of the virtual cluster.

newVirtualCluster :: VirtualCluster

Create a value of VirtualCluster with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:state:VirtualCluster', virtualCluster_state - The state of the virtual cluster.

$sel:arn:VirtualCluster', virtualCluster_arn - The ARN of the virtual cluster.

$sel:createdAt:VirtualCluster', virtualCluster_createdAt - The date and time when the virtual cluster is created.

$sel:name:VirtualCluster', virtualCluster_name - The name of the virtual cluster.

$sel:id:VirtualCluster', virtualCluster_id - The ID of the virtual cluster.

$sel:containerProvider:VirtualCluster', virtualCluster_containerProvider - The container provider of the virtual cluster.

$sel:tags:VirtualCluster', virtualCluster_tags - The assigned tags of the virtual cluster.

virtualCluster_state :: Lens' VirtualCluster (Maybe VirtualClusterState)

The state of the virtual cluster.

virtualCluster_arn :: Lens' VirtualCluster (Maybe Text)

The ARN of the virtual cluster.

virtualCluster_createdAt :: Lens' VirtualCluster (Maybe UTCTime)

The date and time when the virtual cluster is created.

virtualCluster_name :: Lens' VirtualCluster (Maybe Text)

The name of the virtual cluster.

virtualCluster_id :: Lens' VirtualCluster (Maybe Text)

The ID of the virtual cluster.

virtualCluster_containerProvider :: Lens' VirtualCluster (Maybe ContainerProvider)

The container provider of the virtual cluster.

virtualCluster_tags :: Lens' VirtualCluster (Maybe (HashMap Text Text))

The assigned tags of the virtual cluster.