Amazonka.ElastiCache.CreateCacheParameterGroup

Description

Creates a new Amazon ElastiCache cache parameter group. An ElastiCache cache parameter group is a collection of parameters and their values that are applied to all of the nodes in any cluster or replication group using the CacheParameterGroup.

A newly created CacheParameterGroup is an exact duplicate of the default parameter group for the CacheParameterGroupFamily. To customize the newly created CacheParameterGroup you can change the values of specific parameters. For more information, see:

ModifyCacheParameterGroup in the ElastiCache API Reference.

Parameters and Parameter Groups in the ElastiCache User Guide.

Creating a Request

data CreateCacheParameterGroup

Represents the input of a CreateCacheParameterGroup operation.

See: newCreateCacheParameterGroup smart constructor.

Constructors

CreateCacheParameterGroup'

Fields

tags :: Maybe [Tag]

A list of tags to be added to this resource. A tag is a key-value pair. A tag key must be accompanied by a tag value, although null is accepted.

cacheParameterGroupName :: Text

A user-specified name for the cache parameter group.

cacheParameterGroupFamily :: Text

The name of the cache parameter group family that the cache parameter group can be used with.

Valid values are: memcached1.4 | memcached1.5 | memcached1.6 | redis2.6 | redis2.8 | redis3.2 | redis4.0 | redis5.0 | redis6.x |

description :: Text

A user-specified description for the cache parameter group.

newCreateCacheParameterGroup

Arguments

:: Text
$sel:cacheParameterGroupName:CreateCacheParameterGroup'

-> Text
$sel:cacheParameterGroupFamily:CreateCacheParameterGroup'

-> Text
$sel:description:CreateCacheParameterGroup'

-> CreateCacheParameterGroup

Create a value of CreateCacheParameterGroup with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:tags:CreateCacheParameterGroup', createCacheParameterGroup_tags - A list of tags to be added to this resource. A tag is a key-value pair. A tag key must be accompanied by a tag value, although null is accepted.

$sel:cacheParameterGroupName:CreateCacheParameterGroup', createCacheParameterGroup_cacheParameterGroupName - A user-specified name for the cache parameter group.

$sel:cacheParameterGroupFamily:CreateCacheParameterGroup', createCacheParameterGroup_cacheParameterGroupFamily - The name of the cache parameter group family that the cache parameter group can be used with.

Valid values are: memcached1.4 | memcached1.5 | memcached1.6 | redis2.6 | redis2.8 | redis3.2 | redis4.0 | redis5.0 | redis6.x |

$sel:description:CreateCacheParameterGroup', createCacheParameterGroup_description - A user-specified description for the cache parameter group.

Request Lenses

createCacheParameterGroup_tags :: Lens' CreateCacheParameterGroup (Maybe [Tag])

A list of tags to be added to this resource. A tag is a key-value pair. A tag key must be accompanied by a tag value, although null is accepted.

createCacheParameterGroup_cacheParameterGroupName :: Lens' CreateCacheParameterGroup Text

A user-specified name for the cache parameter group.

createCacheParameterGroup_cacheParameterGroupFamily :: Lens' CreateCacheParameterGroup Text

The name of the cache parameter group family that the cache parameter group can be used with.

Valid values are: memcached1.4 | memcached1.5 | memcached1.6 | redis2.6 | redis2.8 | redis3.2 | redis4.0 | redis5.0 | redis6.x |

createCacheParameterGroup_description :: Lens' CreateCacheParameterGroup Text

A user-specified description for the cache parameter group.

Destructuring the Response

data CreateCacheParameterGroupResponse

See: newCreateCacheParameterGroupResponse smart constructor.

Constructors

CreateCacheParameterGroupResponse'

Fields

cacheParameterGroup :: Maybe CacheParameterGroup

httpStatus :: Int

The response's http status code.

newCreateCacheParameterGroupResponse

Arguments

:: Int
$sel:httpStatus:CreateCacheParameterGroupResponse'

-> CreateCacheParameterGroupResponse

Create a value of CreateCacheParameterGroupResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:cacheParameterGroup:CreateCacheParameterGroupResponse', createCacheParameterGroupResponse_cacheParameterGroup - Undocumented member.

$sel:httpStatus:CreateCacheParameterGroupResponse', createCacheParameterGroupResponse_httpStatus - The response's http status code.

Response Lenses

createCacheParameterGroupResponse_cacheParameterGroup :: Lens' CreateCacheParameterGroupResponse (Maybe CacheParameterGroup)

Undocumented member.

createCacheParameterGroupResponse_httpStatus :: Lens' CreateCacheParameterGroupResponse Int

The response's http status code.