Amazonka.ElastiCache.CreateCacheSubnetGroup

Description

Creates a new cache subnet group.

Use this parameter only when you are creating a cluster in an Amazon Virtual Private Cloud (Amazon VPC).

Creating a Request

data CreateCacheSubnetGroup

Represents the input of a CreateCacheSubnetGroup operation.

See: newCreateCacheSubnetGroup smart constructor.

Constructors

CreateCacheSubnetGroup'

Fields

tags :: Maybe [Tag]

A list of tags to be added to this resource. A tag is a key-value pair. A tag key must be accompanied by a tag value, although null is accepted.

cacheSubnetGroupName :: Text

A name for the cache subnet group. This value is stored as a lowercase string.

Constraints: Must contain no more than 255 alphanumeric characters or hyphens.

Example: mysubnetgroup

cacheSubnetGroupDescription :: Text

A description for the cache subnet group.

subnetIds :: [Text]

A list of VPC subnet IDs for the cache subnet group.

newCreateCacheSubnetGroup

Arguments

:: Text
$sel:cacheSubnetGroupName:CreateCacheSubnetGroup'

-> Text
$sel:cacheSubnetGroupDescription:CreateCacheSubnetGroup'

-> CreateCacheSubnetGroup

Create a value of CreateCacheSubnetGroup with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:tags:CreateCacheSubnetGroup', createCacheSubnetGroup_tags - A list of tags to be added to this resource. A tag is a key-value pair. A tag key must be accompanied by a tag value, although null is accepted.

$sel:cacheSubnetGroupName:CreateCacheSubnetGroup', createCacheSubnetGroup_cacheSubnetGroupName - A name for the cache subnet group. This value is stored as a lowercase string.

Constraints: Must contain no more than 255 alphanumeric characters or hyphens.

Example: mysubnetgroup

$sel:cacheSubnetGroupDescription:CreateCacheSubnetGroup', createCacheSubnetGroup_cacheSubnetGroupDescription - A description for the cache subnet group.

$sel:subnetIds:CreateCacheSubnetGroup', createCacheSubnetGroup_subnetIds - A list of VPC subnet IDs for the cache subnet group.

Request Lenses

createCacheSubnetGroup_tags :: Lens' CreateCacheSubnetGroup (Maybe [Tag])

A list of tags to be added to this resource. A tag is a key-value pair. A tag key must be accompanied by a tag value, although null is accepted.

createCacheSubnetGroup_cacheSubnetGroupName :: Lens' CreateCacheSubnetGroup Text

A name for the cache subnet group. This value is stored as a lowercase string.

Constraints: Must contain no more than 255 alphanumeric characters or hyphens.

Example: mysubnetgroup

createCacheSubnetGroup_cacheSubnetGroupDescription :: Lens' CreateCacheSubnetGroup Text

A description for the cache subnet group.

createCacheSubnetGroup_subnetIds :: Lens' CreateCacheSubnetGroup [Text]

A list of VPC subnet IDs for the cache subnet group.

Destructuring the Response

data CreateCacheSubnetGroupResponse

See: newCreateCacheSubnetGroupResponse smart constructor.

Constructors

CreateCacheSubnetGroupResponse'

Fields

cacheSubnetGroup :: Maybe CacheSubnetGroup

httpStatus :: Int

The response's http status code.

newCreateCacheSubnetGroupResponse

Arguments

:: Int
$sel:httpStatus:CreateCacheSubnetGroupResponse'

-> CreateCacheSubnetGroupResponse

Create a value of CreateCacheSubnetGroupResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:cacheSubnetGroup:CreateCacheSubnetGroupResponse', createCacheSubnetGroupResponse_cacheSubnetGroup - Undocumented member.

$sel:httpStatus:CreateCacheSubnetGroupResponse', createCacheSubnetGroupResponse_httpStatus - The response's http status code.

Response Lenses

createCacheSubnetGroupResponse_cacheSubnetGroup :: Lens' CreateCacheSubnetGroupResponse (Maybe CacheSubnetGroup)

Undocumented member.

createCacheSubnetGroupResponse_httpStatus :: Lens' CreateCacheSubnetGroupResponse Int

The response's http status code.