Amazonka.ElastiCache.DecreaseReplicaCount

Description

Dynamically decreases the number of replicas in a Redis (cluster mode disabled) replication group or the number of replica nodes in one or more node groups (shards) of a Redis (cluster mode enabled) replication group. This operation is performed with no cluster down time.

Creating a Request

data DecreaseReplicaCount

See: newDecreaseReplicaCount smart constructor.

Constructors

DecreaseReplicaCount'

Fields

newReplicaCount' :: Maybe Int

The number of read replica nodes you want at the completion of this operation. For Redis (cluster mode disabled) replication groups, this is the number of replica nodes in the replication group. For Redis (cluster mode enabled) replication groups, this is the number of replica nodes in each of the replication group's node groups.

The minimum number of replicas in a shard or replication group is:

Redis (cluster mode disabled)

If Multi-AZ is enabled: 1

If Multi-AZ is not enabled: 0

Redis (cluster mode enabled): 0 (though you will not be able to failover to a replica if your primary node fails)

replicaConfiguration :: Maybe [ConfigureShard]

A list of ConfigureShard objects that can be used to configure each shard in a Redis (cluster mode enabled) replication group. The ConfigureShard has three members: NewReplicaCount, NodeGroupId, and PreferredAvailabilityZones.

replicasToRemove :: Maybe [Text]

A list of the node ids to remove from the replication group or node group (shard).

replicationGroupId :: Text

The id of the replication group from which you want to remove replica nodes.

applyImmediately :: Bool

If True, the number of replica nodes is decreased immediately. ApplyImmediately=False is not currently supported.

newDecreaseReplicaCount

Arguments

:: Text
$sel:replicationGroupId:DecreaseReplicaCount'

-> Bool
$sel:applyImmediately:DecreaseReplicaCount'

-> DecreaseReplicaCount

Create a value of DecreaseReplicaCount with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:newReplicaCount':DecreaseReplicaCount', decreaseReplicaCount_newReplicaCount - The number of read replica nodes you want at the completion of this operation. For Redis (cluster mode disabled) replication groups, this is the number of replica nodes in the replication group. For Redis (cluster mode enabled) replication groups, this is the number of replica nodes in each of the replication group's node groups.

The minimum number of replicas in a shard or replication group is:

Redis (cluster mode disabled)

If Multi-AZ is enabled: 1

If Multi-AZ is not enabled: 0

Redis (cluster mode enabled): 0 (though you will not be able to failover to a replica if your primary node fails)

$sel:replicaConfiguration:DecreaseReplicaCount', decreaseReplicaCount_replicaConfiguration - A list of ConfigureShard objects that can be used to configure each shard in a Redis (cluster mode enabled) replication group. The ConfigureShard has three members: NewReplicaCount, NodeGroupId, and PreferredAvailabilityZones.

$sel:replicasToRemove:DecreaseReplicaCount', decreaseReplicaCount_replicasToRemove - A list of the node ids to remove from the replication group or node group (shard).

$sel:replicationGroupId:DecreaseReplicaCount', decreaseReplicaCount_replicationGroupId - The id of the replication group from which you want to remove replica nodes.

$sel:applyImmediately:DecreaseReplicaCount', decreaseReplicaCount_applyImmediately - If True, the number of replica nodes is decreased immediately. ApplyImmediately=False is not currently supported.

Request Lenses

decreaseReplicaCount_newReplicaCount :: Lens' DecreaseReplicaCount (Maybe Int)

The number of read replica nodes you want at the completion of this operation. For Redis (cluster mode disabled) replication groups, this is the number of replica nodes in the replication group. For Redis (cluster mode enabled) replication groups, this is the number of replica nodes in each of the replication group's node groups.

The minimum number of replicas in a shard or replication group is:

Redis (cluster mode disabled)

If Multi-AZ is enabled: 1

If Multi-AZ is not enabled: 0

Redis (cluster mode enabled): 0 (though you will not be able to failover to a replica if your primary node fails)

decreaseReplicaCount_replicaConfiguration :: Lens' DecreaseReplicaCount (Maybe [ConfigureShard])

A list of ConfigureShard objects that can be used to configure each shard in a Redis (cluster mode enabled) replication group. The ConfigureShard has three members: NewReplicaCount, NodeGroupId, and PreferredAvailabilityZones.

decreaseReplicaCount_replicasToRemove :: Lens' DecreaseReplicaCount (Maybe [Text])

A list of the node ids to remove from the replication group or node group (shard).

decreaseReplicaCount_replicationGroupId :: Lens' DecreaseReplicaCount Text

The id of the replication group from which you want to remove replica nodes.

decreaseReplicaCount_applyImmediately :: Lens' DecreaseReplicaCount Bool

If True, the number of replica nodes is decreased immediately. ApplyImmediately=False is not currently supported.

Destructuring the Response

data DecreaseReplicaCountResponse

See: newDecreaseReplicaCountResponse smart constructor.

Constructors

DecreaseReplicaCountResponse'

Fields

replicationGroup :: Maybe ReplicationGroup

httpStatus :: Int

The response's http status code.

newDecreaseReplicaCountResponse

Arguments

:: Int
$sel:httpStatus:DecreaseReplicaCountResponse'

-> DecreaseReplicaCountResponse

Create a value of DecreaseReplicaCountResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:replicationGroup:DecreaseReplicaCountResponse', decreaseReplicaCountResponse_replicationGroup - Undocumented member.

$sel:httpStatus:DecreaseReplicaCountResponse', decreaseReplicaCountResponse_httpStatus - The response's http status code.

Response Lenses

decreaseReplicaCountResponse_replicationGroup :: Lens' DecreaseReplicaCountResponse (Maybe ReplicationGroup)

Undocumented member.

decreaseReplicaCountResponse_httpStatus :: Lens' DecreaseReplicaCountResponse Int

The response's http status code.