Amazonka.ElastiCache.DeleteReplicationGroup

Description

Deletes an existing replication group. By default, this operation deletes the entire replication group, including the primary/primaries and all of the read replicas. If the replication group has only one primary, you can optionally delete only the read replicas, while retaining the primary by setting RetainPrimaryCluster=true.

When you receive a successful response from this operation, Amazon ElastiCache immediately begins deleting the selected resources; you cannot cancel or revert this operation.

This operation is valid for Redis only.

Creating a Request

data DeleteReplicationGroup

Represents the input of a DeleteReplicationGroup operation.

See: newDeleteReplicationGroup smart constructor.

Constructors

DeleteReplicationGroup'

Fields

finalSnapshotIdentifier :: Maybe Text

The name of a final node group (shard) snapshot. ElastiCache creates the snapshot from the primary node in the cluster, rather than one of the replicas; this is to ensure that it captures the freshest data. After the final snapshot is taken, the replication group is immediately deleted.

retainPrimaryCluster :: Maybe Bool

If set to true, all of the read replicas are deleted, but the primary node is retained.

replicationGroupId :: Text

The identifier for the cluster to be deleted. This parameter is not case sensitive.

newDeleteReplicationGroup

Arguments

:: Text
$sel:replicationGroupId:DeleteReplicationGroup'

-> DeleteReplicationGroup

Create a value of DeleteReplicationGroup with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:finalSnapshotIdentifier:DeleteReplicationGroup', deleteReplicationGroup_finalSnapshotIdentifier - The name of a final node group (shard) snapshot. ElastiCache creates the snapshot from the primary node in the cluster, rather than one of the replicas; this is to ensure that it captures the freshest data. After the final snapshot is taken, the replication group is immediately deleted.

$sel:retainPrimaryCluster:DeleteReplicationGroup', deleteReplicationGroup_retainPrimaryCluster - If set to true, all of the read replicas are deleted, but the primary node is retained.

$sel:replicationGroupId:DeleteReplicationGroup', deleteReplicationGroup_replicationGroupId - The identifier for the cluster to be deleted. This parameter is not case sensitive.

Request Lenses

deleteReplicationGroup_finalSnapshotIdentifier :: Lens' DeleteReplicationGroup (Maybe Text)

The name of a final node group (shard) snapshot. ElastiCache creates the snapshot from the primary node in the cluster, rather than one of the replicas; this is to ensure that it captures the freshest data. After the final snapshot is taken, the replication group is immediately deleted.

deleteReplicationGroup_retainPrimaryCluster :: Lens' DeleteReplicationGroup (Maybe Bool)

If set to true, all of the read replicas are deleted, but the primary node is retained.

deleteReplicationGroup_replicationGroupId :: Lens' DeleteReplicationGroup Text

The identifier for the cluster to be deleted. This parameter is not case sensitive.

Destructuring the Response

data DeleteReplicationGroupResponse

See: newDeleteReplicationGroupResponse smart constructor.

Constructors

DeleteReplicationGroupResponse'

Fields

replicationGroup :: Maybe ReplicationGroup

httpStatus :: Int

The response's http status code.

newDeleteReplicationGroupResponse

Arguments

:: Int
$sel:httpStatus:DeleteReplicationGroupResponse'

-> DeleteReplicationGroupResponse

Create a value of DeleteReplicationGroupResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:replicationGroup:DeleteReplicationGroupResponse', deleteReplicationGroupResponse_replicationGroup - Undocumented member.

$sel:httpStatus:DeleteReplicationGroupResponse', deleteReplicationGroupResponse_httpStatus - The response's http status code.

Response Lenses

deleteReplicationGroupResponse_replicationGroup :: Lens' DeleteReplicationGroupResponse (Maybe ReplicationGroup)

Undocumented member.

deleteReplicationGroupResponse_httpStatus :: Lens' DeleteReplicationGroupResponse Int

The response's http status code.