Amazonka.ElastiCache.DescribeCacheEngineVersions

Description

Returns a list of the available cache engines and their versions.

This operation returns paginated results.

Creating a Request

data DescribeCacheEngineVersions

Represents the input of a DescribeCacheEngineVersions operation.

See: newDescribeCacheEngineVersions smart constructor.

Constructors

DescribeCacheEngineVersions'

Fields

engineVersion :: Maybe Text

The cache engine version to return.

Example: 1.4.14

cacheParameterGroupFamily :: Maybe Text

The name of a specific cache parameter group family to return details for.

Valid values are: memcached1.4 | memcached1.5 | memcached1.6 | redis2.6 | redis2.8 | redis3.2 | redis4.0 | redis5.0 | redis6.x |

Constraints:

Must be 1 to 255 alphanumeric characters

First character must be a letter

Cannot end with a hyphen or contain two consecutive hyphens

defaultOnly :: Maybe Bool

If true, specifies that only the default version of the specified engine or engine and major version combination is to be returned.

engine :: Maybe Text

The cache engine to return. Valid values: memcached | redis

marker :: Maybe Text

An optional marker returned from a prior request. Use this marker for pagination of results from this operation. If this parameter is specified, the response includes only records beyond the marker, up to the value specified by MaxRecords.

maxRecords :: Maybe Int

The maximum number of records to include in the response. If more records exist than the specified MaxRecords value, a marker is included in the response so that the remaining results can be retrieved.

Default: 100

Constraints: minimum 20; maximum 100.

newDescribeCacheEngineVersions :: DescribeCacheEngineVersions

Create a value of DescribeCacheEngineVersions with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:engineVersion:DescribeCacheEngineVersions', describeCacheEngineVersions_engineVersion - The cache engine version to return.

Example: 1.4.14

$sel:cacheParameterGroupFamily:DescribeCacheEngineVersions', describeCacheEngineVersions_cacheParameterGroupFamily - The name of a specific cache parameter group family to return details for.

Valid values are: memcached1.4 | memcached1.5 | memcached1.6 | redis2.6 | redis2.8 | redis3.2 | redis4.0 | redis5.0 | redis6.x |

Constraints:

Must be 1 to 255 alphanumeric characters

First character must be a letter

Cannot end with a hyphen or contain two consecutive hyphens

$sel:defaultOnly:DescribeCacheEngineVersions', describeCacheEngineVersions_defaultOnly - If true, specifies that only the default version of the specified engine or engine and major version combination is to be returned.

$sel:engine:DescribeCacheEngineVersions', describeCacheEngineVersions_engine - The cache engine to return. Valid values: memcached | redis

$sel:marker:DescribeCacheEngineVersions', describeCacheEngineVersions_marker - An optional marker returned from a prior request. Use this marker for pagination of results from this operation. If this parameter is specified, the response includes only records beyond the marker, up to the value specified by MaxRecords.

$sel:maxRecords:DescribeCacheEngineVersions', describeCacheEngineVersions_maxRecords - The maximum number of records to include in the response. If more records exist than the specified MaxRecords value, a marker is included in the response so that the remaining results can be retrieved.

Default: 100

Constraints: minimum 20; maximum 100.

Request Lenses

describeCacheEngineVersions_engineVersion :: Lens' DescribeCacheEngineVersions (Maybe Text)

The cache engine version to return.

Example: 1.4.14

describeCacheEngineVersions_cacheParameterGroupFamily :: Lens' DescribeCacheEngineVersions (Maybe Text)

The name of a specific cache parameter group family to return details for.

Valid values are: memcached1.4 | memcached1.5 | memcached1.6 | redis2.6 | redis2.8 | redis3.2 | redis4.0 | redis5.0 | redis6.x |

Constraints:

Must be 1 to 255 alphanumeric characters

First character must be a letter

Cannot end with a hyphen or contain two consecutive hyphens

describeCacheEngineVersions_defaultOnly :: Lens' DescribeCacheEngineVersions (Maybe Bool)

If true, specifies that only the default version of the specified engine or engine and major version combination is to be returned.

describeCacheEngineVersions_engine :: Lens' DescribeCacheEngineVersions (Maybe Text)

The cache engine to return. Valid values: memcached | redis

describeCacheEngineVersions_marker :: Lens' DescribeCacheEngineVersions (Maybe Text)

An optional marker returned from a prior request. Use this marker for pagination of results from this operation. If this parameter is specified, the response includes only records beyond the marker, up to the value specified by MaxRecords.

describeCacheEngineVersions_maxRecords :: Lens' DescribeCacheEngineVersions (Maybe Int)

The maximum number of records to include in the response. If more records exist than the specified MaxRecords value, a marker is included in the response so that the remaining results can be retrieved.

Default: 100

Constraints: minimum 20; maximum 100.

Destructuring the Response

data DescribeCacheEngineVersionsResponse

Represents the output of a DescribeCacheEngineVersions operation.

See: newDescribeCacheEngineVersionsResponse smart constructor.

Constructors

DescribeCacheEngineVersionsResponse'

Fields

cacheEngineVersions :: Maybe [CacheEngineVersion]

A list of cache engine version details. Each element in the list contains detailed information about one cache engine version.

marker :: Maybe Text

Provides an identifier to allow retrieval of paginated results.

httpStatus :: Int

The response's http status code.

newDescribeCacheEngineVersionsResponse

Arguments

:: Int
$sel:httpStatus:DescribeCacheEngineVersionsResponse'

-> DescribeCacheEngineVersionsResponse

Create a value of DescribeCacheEngineVersionsResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:cacheEngineVersions:DescribeCacheEngineVersionsResponse', describeCacheEngineVersionsResponse_cacheEngineVersions - A list of cache engine version details. Each element in the list contains detailed information about one cache engine version.

$sel:marker:DescribeCacheEngineVersions', describeCacheEngineVersionsResponse_marker - Provides an identifier to allow retrieval of paginated results.

$sel:httpStatus:DescribeCacheEngineVersionsResponse', describeCacheEngineVersionsResponse_httpStatus - The response's http status code.

Response Lenses

describeCacheEngineVersionsResponse_cacheEngineVersions :: Lens' DescribeCacheEngineVersionsResponse (Maybe [CacheEngineVersion])

A list of cache engine version details. Each element in the list contains detailed information about one cache engine version.

describeCacheEngineVersionsResponse_marker :: Lens' DescribeCacheEngineVersionsResponse (Maybe Text)

Provides an identifier to allow retrieval of paginated results.

describeCacheEngineVersionsResponse_httpStatus :: Lens' DescribeCacheEngineVersionsResponse Int

The response's http status code.