Amazonka.ElastiCache.DescribeSnapshots

Description

Returns information about cluster or replication group snapshots. By default, DescribeSnapshots lists all of your snapshots; it can optionally describe a single snapshot, or just the snapshots associated with a particular cache cluster.

This operation is valid for Redis only.

This operation returns paginated results.

Creating a Request

data DescribeSnapshots

Represents the input of a DescribeSnapshotsMessage operation.

See: newDescribeSnapshots smart constructor.

Constructors

DescribeSnapshots'

Fields

cacheClusterId :: Maybe Text

A user-supplied cluster identifier. If this parameter is specified, only snapshots associated with that specific cluster are described.

marker :: Maybe Text

An optional marker returned from a prior request. Use this marker for pagination of results from this operation. If this parameter is specified, the response includes only records beyond the marker, up to the value specified by MaxRecords.

maxRecords :: Maybe Int

The maximum number of records to include in the response. If more records exist than the specified MaxRecords value, a marker is included in the response so that the remaining results can be retrieved.

Default: 50

Constraints: minimum 20; maximum 50.

snapshotName :: Maybe Text

A user-supplied name of the snapshot. If this parameter is specified, only this snapshot are described.

showNodeGroupConfig :: Maybe Bool

A Boolean value which if true, the node group (shard) configuration is included in the snapshot description.

replicationGroupId :: Maybe Text

A user-supplied replication group identifier. If this parameter is specified, only snapshots associated with that specific replication group are described.

snapshotSource :: Maybe Text

If set to system, the output shows snapshots that were automatically created by ElastiCache. If set to user the output shows snapshots that were manually created. If omitted, the output shows both automatically and manually created snapshots.

newDescribeSnapshots :: DescribeSnapshots

Create a value of DescribeSnapshots with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:cacheClusterId:DescribeSnapshots', describeSnapshots_cacheClusterId - A user-supplied cluster identifier. If this parameter is specified, only snapshots associated with that specific cluster are described.

$sel:marker:DescribeSnapshots', describeSnapshots_marker - An optional marker returned from a prior request. Use this marker for pagination of results from this operation. If this parameter is specified, the response includes only records beyond the marker, up to the value specified by MaxRecords.

$sel:maxRecords:DescribeSnapshots', describeSnapshots_maxRecords - The maximum number of records to include in the response. If more records exist than the specified MaxRecords value, a marker is included in the response so that the remaining results can be retrieved.

Default: 50

Constraints: minimum 20; maximum 50.

$sel:snapshotName:DescribeSnapshots', describeSnapshots_snapshotName - A user-supplied name of the snapshot. If this parameter is specified, only this snapshot are described.

$sel:showNodeGroupConfig:DescribeSnapshots', describeSnapshots_showNodeGroupConfig - A Boolean value which if true, the node group (shard) configuration is included in the snapshot description.

$sel:replicationGroupId:DescribeSnapshots', describeSnapshots_replicationGroupId - A user-supplied replication group identifier. If this parameter is specified, only snapshots associated with that specific replication group are described.

$sel:snapshotSource:DescribeSnapshots', describeSnapshots_snapshotSource - If set to system, the output shows snapshots that were automatically created by ElastiCache. If set to user the output shows snapshots that were manually created. If omitted, the output shows both automatically and manually created snapshots.

Request Lenses

describeSnapshots_cacheClusterId :: Lens' DescribeSnapshots (Maybe Text)

A user-supplied cluster identifier. If this parameter is specified, only snapshots associated with that specific cluster are described.

describeSnapshots_marker :: Lens' DescribeSnapshots (Maybe Text)

An optional marker returned from a prior request. Use this marker for pagination of results from this operation. If this parameter is specified, the response includes only records beyond the marker, up to the value specified by MaxRecords.

describeSnapshots_maxRecords :: Lens' DescribeSnapshots (Maybe Int)

The maximum number of records to include in the response. If more records exist than the specified MaxRecords value, a marker is included in the response so that the remaining results can be retrieved.

Default: 50

Constraints: minimum 20; maximum 50.

describeSnapshots_snapshotName :: Lens' DescribeSnapshots (Maybe Text)

A user-supplied name of the snapshot. If this parameter is specified, only this snapshot are described.

describeSnapshots_showNodeGroupConfig :: Lens' DescribeSnapshots (Maybe Bool)

A Boolean value which if true, the node group (shard) configuration is included in the snapshot description.

describeSnapshots_replicationGroupId :: Lens' DescribeSnapshots (Maybe Text)

A user-supplied replication group identifier. If this parameter is specified, only snapshots associated with that specific replication group are described.

describeSnapshots_snapshotSource :: Lens' DescribeSnapshots (Maybe Text)

If set to system, the output shows snapshots that were automatically created by ElastiCache. If set to user the output shows snapshots that were manually created. If omitted, the output shows both automatically and manually created snapshots.

Destructuring the Response

data DescribeSnapshotsResponse

Represents the output of a DescribeSnapshots operation.

See: newDescribeSnapshotsResponse smart constructor.

Constructors

DescribeSnapshotsResponse'

Fields

snapshots :: Maybe [Snapshot]

A list of snapshots. Each item in the list contains detailed information about one snapshot.

marker :: Maybe Text

An optional marker returned from a prior request. Use this marker for pagination of results from this operation. If this parameter is specified, the response includes only records beyond the marker, up to the value specified by MaxRecords.

httpStatus :: Int

The response's http status code.

newDescribeSnapshotsResponse

Arguments

:: Int
$sel:httpStatus:DescribeSnapshotsResponse'

-> DescribeSnapshotsResponse

Create a value of DescribeSnapshotsResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:snapshots:DescribeSnapshotsResponse', describeSnapshotsResponse_snapshots - A list of snapshots. Each item in the list contains detailed information about one snapshot.

$sel:marker:DescribeSnapshots', describeSnapshotsResponse_marker - An optional marker returned from a prior request. Use this marker for pagination of results from this operation. If this parameter is specified, the response includes only records beyond the marker, up to the value specified by MaxRecords.

$sel:httpStatus:DescribeSnapshotsResponse', describeSnapshotsResponse_httpStatus - The response's http status code.

Response Lenses

describeSnapshotsResponse_snapshots :: Lens' DescribeSnapshotsResponse (Maybe [Snapshot])

A list of snapshots. Each item in the list contains detailed information about one snapshot.

describeSnapshotsResponse_marker :: Lens' DescribeSnapshotsResponse (Maybe Text)

An optional marker returned from a prior request. Use this marker for pagination of results from this operation. If this parameter is specified, the response includes only records beyond the marker, up to the value specified by MaxRecords.

describeSnapshotsResponse_httpStatus :: Lens' DescribeSnapshotsResponse Int

The response's http status code.