Amazonka.ElastiCache.FailoverGlobalReplicationGroup

Description

Used to failover the primary region to a selected secondary region. The selected secondary region will become primary, and all other clusters will become secondary.

Creating a Request

data FailoverGlobalReplicationGroup

See: newFailoverGlobalReplicationGroup smart constructor.

Constructors

FailoverGlobalReplicationGroup'

Fields

globalReplicationGroupId :: Text

The name of the Global datastore

primaryRegion :: Text

The Amazon region of the primary cluster of the Global datastore

primaryReplicationGroupId :: Text

The name of the primary replication group

newFailoverGlobalReplicationGroup

Arguments

:: Text
$sel:globalReplicationGroupId:FailoverGlobalReplicationGroup'

-> Text
$sel:primaryRegion:FailoverGlobalReplicationGroup'

-> Text
$sel:primaryReplicationGroupId:FailoverGlobalReplicationGroup'

-> FailoverGlobalReplicationGroup

Create a value of FailoverGlobalReplicationGroup with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:globalReplicationGroupId:FailoverGlobalReplicationGroup', failoverGlobalReplicationGroup_globalReplicationGroupId - The name of the Global datastore

$sel:primaryRegion:FailoverGlobalReplicationGroup', failoverGlobalReplicationGroup_primaryRegion - The Amazon region of the primary cluster of the Global datastore

$sel:primaryReplicationGroupId:FailoverGlobalReplicationGroup', failoverGlobalReplicationGroup_primaryReplicationGroupId - The name of the primary replication group

Request Lenses

failoverGlobalReplicationGroup_globalReplicationGroupId :: Lens' FailoverGlobalReplicationGroup Text

The name of the Global datastore

failoverGlobalReplicationGroup_primaryRegion :: Lens' FailoverGlobalReplicationGroup Text

The Amazon region of the primary cluster of the Global datastore

failoverGlobalReplicationGroup_primaryReplicationGroupId :: Lens' FailoverGlobalReplicationGroup Text

The name of the primary replication group

Destructuring the Response

data FailoverGlobalReplicationGroupResponse

See: newFailoverGlobalReplicationGroupResponse smart constructor.

Constructors

FailoverGlobalReplicationGroupResponse'

Fields

globalReplicationGroup :: Maybe GlobalReplicationGroup

httpStatus :: Int

The response's http status code.

newFailoverGlobalReplicationGroupResponse

Arguments

:: Int
$sel:httpStatus:FailoverGlobalReplicationGroupResponse'

-> FailoverGlobalReplicationGroupResponse

Create a value of FailoverGlobalReplicationGroupResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:globalReplicationGroup:FailoverGlobalReplicationGroupResponse', failoverGlobalReplicationGroupResponse_globalReplicationGroup - Undocumented member.

$sel:httpStatus:FailoverGlobalReplicationGroupResponse', failoverGlobalReplicationGroupResponse_httpStatus - The response's http status code.

Response Lenses

failoverGlobalReplicationGroupResponse_globalReplicationGroup :: Lens' FailoverGlobalReplicationGroupResponse (Maybe GlobalReplicationGroup)

Undocumented member.

failoverGlobalReplicationGroupResponse_httpStatus :: Lens' FailoverGlobalReplicationGroupResponse Int

The response's http status code.