Amazonka.ElastiCache.IncreaseReplicaCount

Description

Dynamically increases the number of replicas in a Redis (cluster mode disabled) replication group or the number of replica nodes in one or more node groups (shards) of a Redis (cluster mode enabled) replication group. This operation is performed with no cluster down time.

Creating a Request

data IncreaseReplicaCount

See: newIncreaseReplicaCount smart constructor.

Constructors

IncreaseReplicaCount'

Fields

newReplicaCount' :: Maybe Int

The number of read replica nodes you want at the completion of this operation. For Redis (cluster mode disabled) replication groups, this is the number of replica nodes in the replication group. For Redis (cluster mode enabled) replication groups, this is the number of replica nodes in each of the replication group's node groups.

replicaConfiguration :: Maybe [ConfigureShard]

A list of ConfigureShard objects that can be used to configure each shard in a Redis (cluster mode enabled) replication group. The ConfigureShard has three members: NewReplicaCount, NodeGroupId, and PreferredAvailabilityZones.

replicationGroupId :: Text

The id of the replication group to which you want to add replica nodes.

applyImmediately :: Bool

If True, the number of replica nodes is increased immediately. ApplyImmediately=False is not currently supported.

newIncreaseReplicaCount

Arguments

:: Text
$sel:replicationGroupId:IncreaseReplicaCount'

-> Bool
$sel:applyImmediately:IncreaseReplicaCount'

-> IncreaseReplicaCount

Create a value of IncreaseReplicaCount with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:newReplicaCount':IncreaseReplicaCount', increaseReplicaCount_newReplicaCount - The number of read replica nodes you want at the completion of this operation. For Redis (cluster mode disabled) replication groups, this is the number of replica nodes in the replication group. For Redis (cluster mode enabled) replication groups, this is the number of replica nodes in each of the replication group's node groups.

$sel:replicaConfiguration:IncreaseReplicaCount', increaseReplicaCount_replicaConfiguration - A list of ConfigureShard objects that can be used to configure each shard in a Redis (cluster mode enabled) replication group. The ConfigureShard has three members: NewReplicaCount, NodeGroupId, and PreferredAvailabilityZones.

$sel:replicationGroupId:IncreaseReplicaCount', increaseReplicaCount_replicationGroupId - The id of the replication group to which you want to add replica nodes.

$sel:applyImmediately:IncreaseReplicaCount', increaseReplicaCount_applyImmediately - If True, the number of replica nodes is increased immediately. ApplyImmediately=False is not currently supported.

Request Lenses

increaseReplicaCount_newReplicaCount :: Lens' IncreaseReplicaCount (Maybe Int)

The number of read replica nodes you want at the completion of this operation. For Redis (cluster mode disabled) replication groups, this is the number of replica nodes in the replication group. For Redis (cluster mode enabled) replication groups, this is the number of replica nodes in each of the replication group's node groups.

increaseReplicaCount_replicaConfiguration :: Lens' IncreaseReplicaCount (Maybe [ConfigureShard])

A list of ConfigureShard objects that can be used to configure each shard in a Redis (cluster mode enabled) replication group. The ConfigureShard has three members: NewReplicaCount, NodeGroupId, and PreferredAvailabilityZones.

increaseReplicaCount_replicationGroupId :: Lens' IncreaseReplicaCount Text

The id of the replication group to which you want to add replica nodes.

increaseReplicaCount_applyImmediately :: Lens' IncreaseReplicaCount Bool

If True, the number of replica nodes is increased immediately. ApplyImmediately=False is not currently supported.

Destructuring the Response

data IncreaseReplicaCountResponse

See: newIncreaseReplicaCountResponse smart constructor.

Constructors

IncreaseReplicaCountResponse'

Fields

replicationGroup :: Maybe ReplicationGroup

httpStatus :: Int

The response's http status code.

newIncreaseReplicaCountResponse

Arguments

:: Int
$sel:httpStatus:IncreaseReplicaCountResponse'

-> IncreaseReplicaCountResponse

Create a value of IncreaseReplicaCountResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:replicationGroup:IncreaseReplicaCountResponse', increaseReplicaCountResponse_replicationGroup - Undocumented member.

$sel:httpStatus:IncreaseReplicaCountResponse', increaseReplicaCountResponse_httpStatus - The response's http status code.

Response Lenses

increaseReplicaCountResponse_replicationGroup :: Lens' IncreaseReplicaCountResponse (Maybe ReplicationGroup)

Undocumented member.

increaseReplicaCountResponse_httpStatus :: Lens' IncreaseReplicaCountResponse Int

The response's http status code.