Amazonka.ElastiCache.ModifyReplicationGroupShardConfiguration

Description

Modifies a replication group's shards (node groups) by allowing you to add shards, remove shards, or rebalance the keyspaces among existing shards.

Creating a Request

data ModifyReplicationGroupShardConfiguration

Represents the input for a ModifyReplicationGroupShardConfiguration operation.

See: newModifyReplicationGroupShardConfiguration smart constructor.

Constructors

ModifyReplicationGroupShardConfiguration'

Fields

nodeGroupsToRetain :: Maybe [Text]

If the value of NodeGroupCount is less than the current number of node groups (shards), then either NodeGroupsToRemove or NodeGroupsToRetain is required. NodeGroupsToRetain is a list of NodeGroupIds to retain in the cluster.

ElastiCache for Redis will attempt to remove all node groups except those listed by NodeGroupsToRetain from the cluster.

reshardingConfiguration :: Maybe [ReshardingConfiguration]

Specifies the preferred availability zones for each node group in the cluster. If the value of NodeGroupCount is greater than the current number of node groups (shards), you can use this parameter to specify the preferred availability zones of the cluster's shards. If you omit this parameter ElastiCache selects availability zones for you.

You can specify this parameter only if the value of NodeGroupCount is greater than the current number of node groups (shards).

nodeGroupsToRemove :: Maybe [Text]

If the value of NodeGroupCount is less than the current number of node groups (shards), then either NodeGroupsToRemove or NodeGroupsToRetain is required. NodeGroupsToRemove is a list of NodeGroupIds to remove from the cluster.

ElastiCache for Redis will attempt to remove all node groups listed by NodeGroupsToRemove from the cluster.

replicationGroupId :: Text

The name of the Redis (cluster mode enabled) cluster (replication group) on which the shards are to be configured.

nodeGroupCount :: Int

The number of node groups (shards) that results from the modification of the shard configuration.

applyImmediately :: Bool

Indicates that the shard reconfiguration process begins immediately. At present, the only permitted value for this parameter is true.

Value: true

newModifyReplicationGroupShardConfiguration

Arguments

:: Text
$sel:replicationGroupId:ModifyReplicationGroupShardConfiguration'

-> Int
$sel:nodeGroupCount:ModifyReplicationGroupShardConfiguration'

-> Bool
$sel:applyImmediately:ModifyReplicationGroupShardConfiguration'

-> ModifyReplicationGroupShardConfiguration

Create a value of ModifyReplicationGroupShardConfiguration with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:nodeGroupsToRetain:ModifyReplicationGroupShardConfiguration', modifyReplicationGroupShardConfiguration_nodeGroupsToRetain - If the value of NodeGroupCount is less than the current number of node groups (shards), then either NodeGroupsToRemove or NodeGroupsToRetain is required. NodeGroupsToRetain is a list of NodeGroupIds to retain in the cluster.

ElastiCache for Redis will attempt to remove all node groups except those listed by NodeGroupsToRetain from the cluster.

$sel:reshardingConfiguration:ModifyReplicationGroupShardConfiguration', modifyReplicationGroupShardConfiguration_reshardingConfiguration - Specifies the preferred availability zones for each node group in the cluster. If the value of NodeGroupCount is greater than the current number of node groups (shards), you can use this parameter to specify the preferred availability zones of the cluster's shards. If you omit this parameter ElastiCache selects availability zones for you.

You can specify this parameter only if the value of NodeGroupCount is greater than the current number of node groups (shards).

$sel:nodeGroupsToRemove:ModifyReplicationGroupShardConfiguration', modifyReplicationGroupShardConfiguration_nodeGroupsToRemove - If the value of NodeGroupCount is less than the current number of node groups (shards), then either NodeGroupsToRemove or NodeGroupsToRetain is required. NodeGroupsToRemove is a list of NodeGroupIds to remove from the cluster.

ElastiCache for Redis will attempt to remove all node groups listed by NodeGroupsToRemove from the cluster.

$sel:replicationGroupId:ModifyReplicationGroupShardConfiguration', modifyReplicationGroupShardConfiguration_replicationGroupId - The name of the Redis (cluster mode enabled) cluster (replication group) on which the shards are to be configured.

$sel:nodeGroupCount:ModifyReplicationGroupShardConfiguration', modifyReplicationGroupShardConfiguration_nodeGroupCount - The number of node groups (shards) that results from the modification of the shard configuration.

$sel:applyImmediately:ModifyReplicationGroupShardConfiguration', modifyReplicationGroupShardConfiguration_applyImmediately - Indicates that the shard reconfiguration process begins immediately. At present, the only permitted value for this parameter is true.

Value: true

Request Lenses

modifyReplicationGroupShardConfiguration_nodeGroupsToRetain :: Lens' ModifyReplicationGroupShardConfiguration (Maybe [Text])

If the value of NodeGroupCount is less than the current number of node groups (shards), then either NodeGroupsToRemove or NodeGroupsToRetain is required. NodeGroupsToRetain is a list of NodeGroupIds to retain in the cluster.

ElastiCache for Redis will attempt to remove all node groups except those listed by NodeGroupsToRetain from the cluster.

modifyReplicationGroupShardConfiguration_reshardingConfiguration :: Lens' ModifyReplicationGroupShardConfiguration (Maybe [ReshardingConfiguration])

Specifies the preferred availability zones for each node group in the cluster. If the value of NodeGroupCount is greater than the current number of node groups (shards), you can use this parameter to specify the preferred availability zones of the cluster's shards. If you omit this parameter ElastiCache selects availability zones for you.

You can specify this parameter only if the value of NodeGroupCount is greater than the current number of node groups (shards).

modifyReplicationGroupShardConfiguration_nodeGroupsToRemove :: Lens' ModifyReplicationGroupShardConfiguration (Maybe [Text])

If the value of NodeGroupCount is less than the current number of node groups (shards), then either NodeGroupsToRemove or NodeGroupsToRetain is required. NodeGroupsToRemove is a list of NodeGroupIds to remove from the cluster.

ElastiCache for Redis will attempt to remove all node groups listed by NodeGroupsToRemove from the cluster.

modifyReplicationGroupShardConfiguration_replicationGroupId :: Lens' ModifyReplicationGroupShardConfiguration Text

The name of the Redis (cluster mode enabled) cluster (replication group) on which the shards are to be configured.

modifyReplicationGroupShardConfiguration_nodeGroupCount :: Lens' ModifyReplicationGroupShardConfiguration Int

The number of node groups (shards) that results from the modification of the shard configuration.

modifyReplicationGroupShardConfiguration_applyImmediately :: Lens' ModifyReplicationGroupShardConfiguration Bool

Indicates that the shard reconfiguration process begins immediately. At present, the only permitted value for this parameter is true.

Value: true

Destructuring the Response

data ModifyReplicationGroupShardConfigurationResponse

See: newModifyReplicationGroupShardConfigurationResponse smart constructor.

Constructors

ModifyReplicationGroupShardConfigurationResponse'

Fields

replicationGroup :: Maybe ReplicationGroup

httpStatus :: Int

The response's http status code.

newModifyReplicationGroupShardConfigurationResponse

Arguments

:: Int
$sel:httpStatus:ModifyReplicationGroupShardConfigurationResponse'

-> ModifyReplicationGroupShardConfigurationResponse

Create a value of ModifyReplicationGroupShardConfigurationResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:replicationGroup:ModifyReplicationGroupShardConfigurationResponse', modifyReplicationGroupShardConfigurationResponse_replicationGroup - Undocumented member.

$sel:httpStatus:ModifyReplicationGroupShardConfigurationResponse', modifyReplicationGroupShardConfigurationResponse_httpStatus - The response's http status code.

Response Lenses

modifyReplicationGroupShardConfigurationResponse_replicationGroup :: Lens' ModifyReplicationGroupShardConfigurationResponse (Maybe ReplicationGroup)

Undocumented member.

modifyReplicationGroupShardConfigurationResponse_httpStatus :: Lens' ModifyReplicationGroupShardConfigurationResponse Int

The response's http status code.