Amazonka.ElastiCache.PurchaseReservedCacheNodesOffering

Description

Allows you to purchase a reserved cache node offering. Reserved nodes are not eligible for cancellation and are non-refundable. For more information, see Managing Costs with Reserved Nodes for Redis or Managing Costs with Reserved Nodes for Memcached.

Creating a Request

data PurchaseReservedCacheNodesOffering

Represents the input of a PurchaseReservedCacheNodesOffering operation.

See: newPurchaseReservedCacheNodesOffering smart constructor.

Constructors

PurchaseReservedCacheNodesOffering'

Fields

cacheNodeCount :: Maybe Int

The number of cache node instances to reserve.

Default: 1

reservedCacheNodeId :: Maybe Text

A customer-specified identifier to track this reservation.

The Reserved Cache Node ID is an unique customer-specified identifier to track this reservation. If this parameter is not specified, ElastiCache automatically generates an identifier for the reservation.

Example: myreservationID

tags :: Maybe [Tag]

A list of tags to be added to this resource. A tag is a key-value pair. A tag key must be accompanied by a tag value, although null is accepted.

reservedCacheNodesOfferingId :: Text

The ID of the reserved cache node offering to purchase.

Example: 438012d3-4052-4cc7-b2e3-8d3372e0e706

newPurchaseReservedCacheNodesOffering

Arguments

:: Text
$sel:reservedCacheNodesOfferingId:PurchaseReservedCacheNodesOffering'

-> PurchaseReservedCacheNodesOffering

Create a value of PurchaseReservedCacheNodesOffering with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:cacheNodeCount:PurchaseReservedCacheNodesOffering', purchaseReservedCacheNodesOffering_cacheNodeCount - The number of cache node instances to reserve.

Default: 1

$sel:reservedCacheNodeId:PurchaseReservedCacheNodesOffering', purchaseReservedCacheNodesOffering_reservedCacheNodeId - A customer-specified identifier to track this reservation.

The Reserved Cache Node ID is an unique customer-specified identifier to track this reservation. If this parameter is not specified, ElastiCache automatically generates an identifier for the reservation.

Example: myreservationID

$sel:tags:PurchaseReservedCacheNodesOffering', purchaseReservedCacheNodesOffering_tags - A list of tags to be added to this resource. A tag is a key-value pair. A tag key must be accompanied by a tag value, although null is accepted.

$sel:reservedCacheNodesOfferingId:PurchaseReservedCacheNodesOffering', purchaseReservedCacheNodesOffering_reservedCacheNodesOfferingId - The ID of the reserved cache node offering to purchase.

Example: 438012d3-4052-4cc7-b2e3-8d3372e0e706

Request Lenses

purchaseReservedCacheNodesOffering_cacheNodeCount :: Lens' PurchaseReservedCacheNodesOffering (Maybe Int)

The number of cache node instances to reserve.

Default: 1

purchaseReservedCacheNodesOffering_reservedCacheNodeId :: Lens' PurchaseReservedCacheNodesOffering (Maybe Text)

A customer-specified identifier to track this reservation.

The Reserved Cache Node ID is an unique customer-specified identifier to track this reservation. If this parameter is not specified, ElastiCache automatically generates an identifier for the reservation.

Example: myreservationID

purchaseReservedCacheNodesOffering_tags :: Lens' PurchaseReservedCacheNodesOffering (Maybe [Tag])

A list of tags to be added to this resource. A tag is a key-value pair. A tag key must be accompanied by a tag value, although null is accepted.

purchaseReservedCacheNodesOffering_reservedCacheNodesOfferingId :: Lens' PurchaseReservedCacheNodesOffering Text

The ID of the reserved cache node offering to purchase.

Example: 438012d3-4052-4cc7-b2e3-8d3372e0e706

Destructuring the Response

data PurchaseReservedCacheNodesOfferingResponse

See: newPurchaseReservedCacheNodesOfferingResponse smart constructor.

Constructors

PurchaseReservedCacheNodesOfferingResponse'

Fields

reservedCacheNode :: Maybe ReservedCacheNode

httpStatus :: Int

The response's http status code.

newPurchaseReservedCacheNodesOfferingResponse

Arguments

:: Int
$sel:httpStatus:PurchaseReservedCacheNodesOfferingResponse'

-> PurchaseReservedCacheNodesOfferingResponse

Create a value of PurchaseReservedCacheNodesOfferingResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:reservedCacheNode:PurchaseReservedCacheNodesOfferingResponse', purchaseReservedCacheNodesOfferingResponse_reservedCacheNode - Undocumented member.

$sel:httpStatus:PurchaseReservedCacheNodesOfferingResponse', purchaseReservedCacheNodesOfferingResponse_httpStatus - The response's http status code.

Response Lenses

purchaseReservedCacheNodesOfferingResponse_reservedCacheNode :: Lens' PurchaseReservedCacheNodesOfferingResponse (Maybe ReservedCacheNode)

Undocumented member.

purchaseReservedCacheNodesOfferingResponse_httpStatus :: Lens' PurchaseReservedCacheNodesOfferingResponse Int

The response's http status code.