Amazonka.ElastiCache.StartMigration

Description

Start the migration of data.

Creating a Request

data StartMigration

See: newStartMigration smart constructor.

Constructors

StartMigration'

Fields

replicationGroupId :: Text

The ID of the replication group to which data should be migrated.

customerNodeEndpointList :: [CustomerNodeEndpoint]

List of endpoints from which data should be migrated. For Redis (cluster mode disabled), list should have only one element.

newStartMigration

Arguments

:: Text
$sel:replicationGroupId:StartMigration'

-> StartMigration

Create a value of StartMigration with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:replicationGroupId:StartMigration', startMigration_replicationGroupId - The ID of the replication group to which data should be migrated.

$sel:customerNodeEndpointList:StartMigration', startMigration_customerNodeEndpointList - List of endpoints from which data should be migrated. For Redis (cluster mode disabled), list should have only one element.

Request Lenses

startMigration_replicationGroupId :: Lens' StartMigration Text

The ID of the replication group to which data should be migrated.

startMigration_customerNodeEndpointList :: Lens' StartMigration [CustomerNodeEndpoint]

List of endpoints from which data should be migrated. For Redis (cluster mode disabled), list should have only one element.

Destructuring the Response

data StartMigrationResponse

See: newStartMigrationResponse smart constructor.

Constructors

StartMigrationResponse'

Fields

replicationGroup :: Maybe ReplicationGroup

httpStatus :: Int

The response's http status code.

newStartMigrationResponse

Arguments

:: Int
$sel:httpStatus:StartMigrationResponse'

-> StartMigrationResponse

Create a value of StartMigrationResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:replicationGroup:StartMigrationResponse', startMigrationResponse_replicationGroup - Undocumented member.

$sel:httpStatus:StartMigrationResponse', startMigrationResponse_httpStatus - The response's http status code.

Response Lenses

startMigrationResponse_replicationGroup :: Lens' StartMigrationResponse (Maybe ReplicationGroup)

Undocumented member.

startMigrationResponse_httpStatus :: Lens' StartMigrationResponse Int

The response's http status code.