Amazonka.ElastiCache.Types.CacheCluster

Description

Documentation

data CacheCluster

Contains all of the attributes of a specific cluster.

See: newCacheCluster smart constructor.

Constructors

CacheCluster'

Fields

authTokenLastModifiedDate :: Maybe ISO8601

The date the auth token was last modified

engineVersion :: Maybe Text

The version of the cache engine that is used in this cluster.

cacheNodeType :: Maybe Text

The name of the compute and memory capacity node type for the cluster.

The following node types are supported by ElastiCache. Generally speaking, the current generation types provide more memory and computational power at lower cost when compared to their equivalent previous generation counterparts.

General purpose:

Current generation:

M6g node types (available only for Redis engine version 5.0.6 onward and for Memcached engine version 1.5.16 onward).

cache.m6g.large, cache.m6g.xlarge, cache.m6g.2xlarge, cache.m6g.4xlarge, cache.m6g.8xlarge, cache.m6g.12xlarge, cache.m6g.16xlarge

For region availability, see Supported Node Types

M5 node types: cache.m5.large, cache.m5.xlarge, cache.m5.2xlarge, cache.m5.4xlarge, cache.m5.12xlarge, cache.m5.24xlarge

M4 node types: cache.m4.large, cache.m4.xlarge, cache.m4.2xlarge, cache.m4.4xlarge, cache.m4.10xlarge

T3 node types: cache.t3.micro, cache.t3.small, cache.t3.medium

T2 node types: cache.t2.micro, cache.t2.small, cache.t2.medium

Previous generation: (not recommended)

T1 node types: cache.t1.micro

M1 node types: cache.m1.small, cache.m1.medium, cache.m1.large, cache.m1.xlarge

M3 node types: cache.m3.medium, cache.m3.large, cache.m3.xlarge, cache.m3.2xlarge

Compute optimized:

Previous generation: (not recommended)

C1 node types: cache.c1.xlarge

Memory optimized:

Current generation:

R6g node types (available only for Redis engine version 5.0.6 onward and for Memcached engine version 1.5.16 onward).

cache.r6g.large, cache.r6g.xlarge, cache.r6g.2xlarge, cache.r6g.4xlarge, cache.r6g.8xlarge, cache.r6g.12xlarge, cache.r6g.16xlarge

For region availability, see Supported Node Types

R5 node types: cache.r5.large, cache.r5.xlarge, cache.r5.2xlarge, cache.r5.4xlarge, cache.r5.12xlarge, cache.r5.24xlarge

R4 node types: cache.r4.large, cache.r4.xlarge, cache.r4.2xlarge, cache.r4.4xlarge, cache.r4.8xlarge, cache.r4.16xlarge

Previous generation: (not recommended)

M2 node types: cache.m2.xlarge, cache.m2.2xlarge, cache.m2.4xlarge

R3 node types: cache.r3.large, cache.r3.xlarge, cache.r3.2xlarge, cache.r3.4xlarge, cache.r3.8xlarge

Additional node type info

All current generation instance types are created in Amazon VPC by default.

Redis append-only files (AOF) are not supported for T1 or T2 instances.

Redis Multi-AZ with automatic failover is not supported on T1 instances.

Redis configuration variables appendonly and appendfsync are not supported on Redis version 2.8.22 and later.

cacheNodes :: Maybe [CacheNode]

A list of cache nodes that are members of the cluster.

cacheClusterCreateTime :: Maybe ISO8601

The date and time when the cluster was created.

atRestEncryptionEnabled :: Maybe Bool

A flag that enables encryption at-rest when set to true.

You cannot modify the value of AtRestEncryptionEnabled after the cluster is created. To enable at-rest encryption on a cluster you must set AtRestEncryptionEnabled to true when you create a cluster.

Required: Only available when creating a replication group in an Amazon VPC using redis version 3.2.6, 4.x or later.

Default: false

autoMinorVersionUpgrade :: Maybe Bool

This parameter is currently disabled.

securityGroups :: Maybe [SecurityGroupMembership]

A list of VPC Security Groups associated with the cluster.

notificationConfiguration :: Maybe NotificationConfiguration

Describes a notification topic and its status. Notification topics are used for publishing ElastiCache events to subscribers using Amazon Simple Notification Service (SNS).

arn :: Maybe Text

The ARN (Amazon Resource Name) of the cache cluster.

transitEncryptionEnabled :: Maybe Bool

A flag that enables in-transit encryption when set to true.

You cannot modify the value of TransitEncryptionEnabled after the cluster is created. To enable in-transit encryption on a cluster you must set TransitEncryptionEnabled to true when you create a cluster.

Required: Only available when creating a replication group in an Amazon VPC using redis version 3.2.6, 4.x or later.

Default: false

snapshotWindow :: Maybe Text

The daily time range (in UTC) during which ElastiCache begins taking a daily snapshot of your cluster.

Example: 05:00-09:00

cacheClusterId :: Maybe Text

The user-supplied identifier of the cluster. This identifier is a unique key that identifies a cluster.

logDeliveryConfigurations :: Maybe [LogDeliveryConfiguration]

Returns the destination, format and type of the logs.

configurationEndpoint :: Maybe Endpoint

Represents a Memcached cluster endpoint which can be used by an application to connect to any node in the cluster. The configuration endpoint will always have .cfg in it.

Example: mem-3.9dvc4r.cfg.usw2.cache.amazonaws.com:11211

engine :: Maybe Text

The name of the cache engine (memcached or redis) to be used for this cluster.

cacheSecurityGroups :: Maybe [CacheSecurityGroupMembership]

A list of cache security group elements, composed of name and status sub-elements.

replicationGroupLogDeliveryEnabled :: Maybe Bool

A boolean value indicating whether log delivery is enabled for the replication group.

authTokenEnabled :: Maybe Bool

A flag that enables using an AuthToken (password) when issuing Redis commands.

Default: false

clientDownloadLandingPage :: Maybe Text

The URL of the web page where you can download the latest ElastiCache client library.

preferredMaintenanceWindow :: Maybe Text

Specifies the weekly time range during which maintenance on the cluster is performed. It is specified as a range in the format ddd:hh24:mi-ddd:hh24:mi (24H Clock UTC). The minimum maintenance window is a 60 minute period.

Valid values for ddd are:

sun

mon

tue

wed

thu

fri

sat

Example: sun:23:00-mon:01:30

cacheSubnetGroupName :: Maybe Text

The name of the cache subnet group associated with the cluster.

preferredAvailabilityZone :: Maybe Text

The name of the Availability Zone in which the cluster is located or "Multiple" if the cache nodes are located in different Availability Zones.

cacheParameterGroup :: Maybe CacheParameterGroupStatus

Status of the cache parameter group.

cacheClusterStatus :: Maybe Text

The current state of this cluster, one of the following values: available, creating, deleted, deleting, incompatible-network, modifying, rebooting cluster nodes, restore-failed, or snapshotting.

snapshotRetentionLimit :: Maybe Int

The number of days for which ElastiCache retains automatic cluster snapshots before deleting them. For example, if you set SnapshotRetentionLimit to 5, a snapshot that was taken today is retained for 5 days before being deleted.

If the value of SnapshotRetentionLimit is set to zero (0), backups are turned off.

preferredOutpostArn :: Maybe Text

The outpost ARN in which the cache cluster is created.

replicationGroupId :: Maybe Text

The replication group to which this cluster belongs. If this field is empty, the cluster is not associated with any replication group.

pendingModifiedValues :: Maybe PendingModifiedValues

numCacheNodes :: Maybe Int

The number of cache nodes in the cluster.

For clusters running Redis, this value must be 1. For clusters running Memcached, this value must be between 1 and 40.

newCacheCluster :: CacheCluster

Create a value of CacheCluster with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:authTokenLastModifiedDate:CacheCluster', cacheCluster_authTokenLastModifiedDate - The date the auth token was last modified

$sel:engineVersion:CacheCluster', cacheCluster_engineVersion - The version of the cache engine that is used in this cluster.

$sel:cacheNodeType:CacheCluster', cacheCluster_cacheNodeType - The name of the compute and memory capacity node type for the cluster.

The following node types are supported by ElastiCache. Generally speaking, the current generation types provide more memory and computational power at lower cost when compared to their equivalent previous generation counterparts.

General purpose:

Current generation:

M6g node types (available only for Redis engine version 5.0.6 onward and for Memcached engine version 1.5.16 onward).

cache.m6g.large, cache.m6g.xlarge, cache.m6g.2xlarge, cache.m6g.4xlarge, cache.m6g.8xlarge, cache.m6g.12xlarge, cache.m6g.16xlarge

For region availability, see Supported Node Types

M5 node types: cache.m5.large, cache.m5.xlarge, cache.m5.2xlarge, cache.m5.4xlarge, cache.m5.12xlarge, cache.m5.24xlarge

M4 node types: cache.m4.large, cache.m4.xlarge, cache.m4.2xlarge, cache.m4.4xlarge, cache.m4.10xlarge

T3 node types: cache.t3.micro, cache.t3.small, cache.t3.medium

T2 node types: cache.t2.micro, cache.t2.small, cache.t2.medium

Previous generation: (not recommended)

T1 node types: cache.t1.micro

M1 node types: cache.m1.small, cache.m1.medium, cache.m1.large, cache.m1.xlarge

M3 node types: cache.m3.medium, cache.m3.large, cache.m3.xlarge, cache.m3.2xlarge

Compute optimized:

Previous generation: (not recommended)

C1 node types: cache.c1.xlarge

Memory optimized:

Current generation:

R6g node types (available only for Redis engine version 5.0.6 onward and for Memcached engine version 1.5.16 onward).

cache.r6g.large, cache.r6g.xlarge, cache.r6g.2xlarge, cache.r6g.4xlarge, cache.r6g.8xlarge, cache.r6g.12xlarge, cache.r6g.16xlarge

For region availability, see Supported Node Types

R5 node types: cache.r5.large, cache.r5.xlarge, cache.r5.2xlarge, cache.r5.4xlarge, cache.r5.12xlarge, cache.r5.24xlarge

R4 node types: cache.r4.large, cache.r4.xlarge, cache.r4.2xlarge, cache.r4.4xlarge, cache.r4.8xlarge, cache.r4.16xlarge

Previous generation: (not recommended)

M2 node types: cache.m2.xlarge, cache.m2.2xlarge, cache.m2.4xlarge

R3 node types: cache.r3.large, cache.r3.xlarge, cache.r3.2xlarge, cache.r3.4xlarge, cache.r3.8xlarge

Additional node type info

All current generation instance types are created in Amazon VPC by default.

Redis append-only files (AOF) are not supported for T1 or T2 instances.

Redis Multi-AZ with automatic failover is not supported on T1 instances.

Redis configuration variables appendonly and appendfsync are not supported on Redis version 2.8.22 and later.

$sel:cacheNodes:CacheCluster', cacheCluster_cacheNodes - A list of cache nodes that are members of the cluster.

$sel:cacheClusterCreateTime:CacheCluster', cacheCluster_cacheClusterCreateTime - The date and time when the cluster was created.

$sel:atRestEncryptionEnabled:CacheCluster', cacheCluster_atRestEncryptionEnabled - A flag that enables encryption at-rest when set to true.

You cannot modify the value of AtRestEncryptionEnabled after the cluster is created. To enable at-rest encryption on a cluster you must set AtRestEncryptionEnabled to true when you create a cluster.

Required: Only available when creating a replication group in an Amazon VPC using redis version 3.2.6, 4.x or later.

Default: false

$sel:autoMinorVersionUpgrade:CacheCluster', cacheCluster_autoMinorVersionUpgrade - This parameter is currently disabled.

$sel:securityGroups:CacheCluster', cacheCluster_securityGroups - A list of VPC Security Groups associated with the cluster.

$sel:notificationConfiguration:CacheCluster', cacheCluster_notificationConfiguration - Describes a notification topic and its status. Notification topics are used for publishing ElastiCache events to subscribers using Amazon Simple Notification Service (SNS).

$sel:arn:CacheCluster', cacheCluster_arn - The ARN (Amazon Resource Name) of the cache cluster.

$sel:transitEncryptionEnabled:CacheCluster', cacheCluster_transitEncryptionEnabled - A flag that enables in-transit encryption when set to true.

You cannot modify the value of TransitEncryptionEnabled after the cluster is created. To enable in-transit encryption on a cluster you must set TransitEncryptionEnabled to true when you create a cluster.

Required: Only available when creating a replication group in an Amazon VPC using redis version 3.2.6, 4.x or later.

Default: false

$sel:snapshotWindow:CacheCluster', cacheCluster_snapshotWindow - The daily time range (in UTC) during which ElastiCache begins taking a daily snapshot of your cluster.

Example: 05:00-09:00

$sel:cacheClusterId:CacheCluster', cacheCluster_cacheClusterId - The user-supplied identifier of the cluster. This identifier is a unique key that identifies a cluster.

$sel:logDeliveryConfigurations:CacheCluster', cacheCluster_logDeliveryConfigurations - Returns the destination, format and type of the logs.

$sel:configurationEndpoint:CacheCluster', cacheCluster_configurationEndpoint - Represents a Memcached cluster endpoint which can be used by an application to connect to any node in the cluster. The configuration endpoint will always have .cfg in it.

Example: mem-3.9dvc4r.cfg.usw2.cache.amazonaws.com:11211

$sel:engine:CacheCluster', cacheCluster_engine - The name of the cache engine (memcached or redis) to be used for this cluster.

$sel:cacheSecurityGroups:CacheCluster', cacheCluster_cacheSecurityGroups - A list of cache security group elements, composed of name and status sub-elements.

$sel:replicationGroupLogDeliveryEnabled:CacheCluster', cacheCluster_replicationGroupLogDeliveryEnabled - A boolean value indicating whether log delivery is enabled for the replication group.

$sel:authTokenEnabled:CacheCluster', cacheCluster_authTokenEnabled - A flag that enables using an AuthToken (password) when issuing Redis commands.

Default: false

$sel:clientDownloadLandingPage:CacheCluster', cacheCluster_clientDownloadLandingPage - The URL of the web page where you can download the latest ElastiCache client library.

$sel:preferredMaintenanceWindow:CacheCluster', cacheCluster_preferredMaintenanceWindow - Specifies the weekly time range during which maintenance on the cluster is performed. It is specified as a range in the format ddd:hh24:mi-ddd:hh24:mi (24H Clock UTC). The minimum maintenance window is a 60 minute period.

Valid values for ddd are:

sun

mon

tue

wed

thu

fri

sat

Example: sun:23:00-mon:01:30

$sel:cacheSubnetGroupName:CacheCluster', cacheCluster_cacheSubnetGroupName - The name of the cache subnet group associated with the cluster.

$sel:preferredAvailabilityZone:CacheCluster', cacheCluster_preferredAvailabilityZone - The name of the Availability Zone in which the cluster is located or "Multiple" if the cache nodes are located in different Availability Zones.

$sel:cacheParameterGroup:CacheCluster', cacheCluster_cacheParameterGroup - Status of the cache parameter group.

$sel:cacheClusterStatus:CacheCluster', cacheCluster_cacheClusterStatus - The current state of this cluster, one of the following values: available, creating, deleted, deleting, incompatible-network, modifying, rebooting cluster nodes, restore-failed, or snapshotting.

$sel:snapshotRetentionLimit:CacheCluster', cacheCluster_snapshotRetentionLimit - The number of days for which ElastiCache retains automatic cluster snapshots before deleting them. For example, if you set SnapshotRetentionLimit to 5, a snapshot that was taken today is retained for 5 days before being deleted.

If the value of SnapshotRetentionLimit is set to zero (0), backups are turned off.

$sel:preferredOutpostArn:CacheCluster', cacheCluster_preferredOutpostArn - The outpost ARN in which the cache cluster is created.

$sel:replicationGroupId:CacheCluster', cacheCluster_replicationGroupId - The replication group to which this cluster belongs. If this field is empty, the cluster is not associated with any replication group.

$sel:pendingModifiedValues:CacheCluster', cacheCluster_pendingModifiedValues - Undocumented member.

$sel:numCacheNodes:CacheCluster', cacheCluster_numCacheNodes - The number of cache nodes in the cluster.

For clusters running Redis, this value must be 1. For clusters running Memcached, this value must be between 1 and 40.

cacheCluster_authTokenLastModifiedDate :: Lens' CacheCluster (Maybe UTCTime)

The date the auth token was last modified

cacheCluster_engineVersion :: Lens' CacheCluster (Maybe Text)

The version of the cache engine that is used in this cluster.

cacheCluster_cacheNodeType :: Lens' CacheCluster (Maybe Text)

The name of the compute and memory capacity node type for the cluster.

The following node types are supported by ElastiCache. Generally speaking, the current generation types provide more memory and computational power at lower cost when compared to their equivalent previous generation counterparts.

General purpose:

Current generation:

M6g node types (available only for Redis engine version 5.0.6 onward and for Memcached engine version 1.5.16 onward).

cache.m6g.large, cache.m6g.xlarge, cache.m6g.2xlarge, cache.m6g.4xlarge, cache.m6g.8xlarge, cache.m6g.12xlarge, cache.m6g.16xlarge

For region availability, see Supported Node Types

M5 node types: cache.m5.large, cache.m5.xlarge, cache.m5.2xlarge, cache.m5.4xlarge, cache.m5.12xlarge, cache.m5.24xlarge

M4 node types: cache.m4.large, cache.m4.xlarge, cache.m4.2xlarge, cache.m4.4xlarge, cache.m4.10xlarge

T3 node types: cache.t3.micro, cache.t3.small, cache.t3.medium

T2 node types: cache.t2.micro, cache.t2.small, cache.t2.medium

Previous generation: (not recommended)

T1 node types: cache.t1.micro

M1 node types: cache.m1.small, cache.m1.medium, cache.m1.large, cache.m1.xlarge

M3 node types: cache.m3.medium, cache.m3.large, cache.m3.xlarge, cache.m3.2xlarge

Compute optimized:

Previous generation: (not recommended)

C1 node types: cache.c1.xlarge

Memory optimized:

Current generation:

R6g node types (available only for Redis engine version 5.0.6 onward and for Memcached engine version 1.5.16 onward).

cache.r6g.large, cache.r6g.xlarge, cache.r6g.2xlarge, cache.r6g.4xlarge, cache.r6g.8xlarge, cache.r6g.12xlarge, cache.r6g.16xlarge

For region availability, see Supported Node Types

R5 node types: cache.r5.large, cache.r5.xlarge, cache.r5.2xlarge, cache.r5.4xlarge, cache.r5.12xlarge, cache.r5.24xlarge

R4 node types: cache.r4.large, cache.r4.xlarge, cache.r4.2xlarge, cache.r4.4xlarge, cache.r4.8xlarge, cache.r4.16xlarge

Previous generation: (not recommended)

M2 node types: cache.m2.xlarge, cache.m2.2xlarge, cache.m2.4xlarge

R3 node types: cache.r3.large, cache.r3.xlarge, cache.r3.2xlarge, cache.r3.4xlarge, cache.r3.8xlarge

Additional node type info

All current generation instance types are created in Amazon VPC by default.

Redis append-only files (AOF) are not supported for T1 or T2 instances.

Redis Multi-AZ with automatic failover is not supported on T1 instances.

Redis configuration variables appendonly and appendfsync are not supported on Redis version 2.8.22 and later.

cacheCluster_cacheNodes :: Lens' CacheCluster (Maybe [CacheNode])

A list of cache nodes that are members of the cluster.

cacheCluster_cacheClusterCreateTime :: Lens' CacheCluster (Maybe UTCTime)

The date and time when the cluster was created.

cacheCluster_atRestEncryptionEnabled :: Lens' CacheCluster (Maybe Bool)

A flag that enables encryption at-rest when set to true.

You cannot modify the value of AtRestEncryptionEnabled after the cluster is created. To enable at-rest encryption on a cluster you must set AtRestEncryptionEnabled to true when you create a cluster.

Required: Only available when creating a replication group in an Amazon VPC using redis version 3.2.6, 4.x or later.

Default: false

cacheCluster_autoMinorVersionUpgrade :: Lens' CacheCluster (Maybe Bool)

This parameter is currently disabled.

cacheCluster_securityGroups :: Lens' CacheCluster (Maybe [SecurityGroupMembership])

A list of VPC Security Groups associated with the cluster.

cacheCluster_notificationConfiguration :: Lens' CacheCluster (Maybe NotificationConfiguration)

Describes a notification topic and its status. Notification topics are used for publishing ElastiCache events to subscribers using Amazon Simple Notification Service (SNS).

cacheCluster_arn :: Lens' CacheCluster (Maybe Text)

The ARN (Amazon Resource Name) of the cache cluster.

cacheCluster_transitEncryptionEnabled :: Lens' CacheCluster (Maybe Bool)

A flag that enables in-transit encryption when set to true.

You cannot modify the value of TransitEncryptionEnabled after the cluster is created. To enable in-transit encryption on a cluster you must set TransitEncryptionEnabled to true when you create a cluster.

Required: Only available when creating a replication group in an Amazon VPC using redis version 3.2.6, 4.x or later.

Default: false

cacheCluster_snapshotWindow :: Lens' CacheCluster (Maybe Text)

The daily time range (in UTC) during which ElastiCache begins taking a daily snapshot of your cluster.

Example: 05:00-09:00

cacheCluster_cacheClusterId :: Lens' CacheCluster (Maybe Text)

The user-supplied identifier of the cluster. This identifier is a unique key that identifies a cluster.

cacheCluster_logDeliveryConfigurations :: Lens' CacheCluster (Maybe [LogDeliveryConfiguration])

Returns the destination, format and type of the logs.

cacheCluster_configurationEndpoint :: Lens' CacheCluster (Maybe Endpoint)

Represents a Memcached cluster endpoint which can be used by an application to connect to any node in the cluster. The configuration endpoint will always have .cfg in it.

Example: mem-3.9dvc4r.cfg.usw2.cache.amazonaws.com:11211

cacheCluster_engine :: Lens' CacheCluster (Maybe Text)

The name of the cache engine (memcached or redis) to be used for this cluster.

cacheCluster_cacheSecurityGroups :: Lens' CacheCluster (Maybe [CacheSecurityGroupMembership])

A list of cache security group elements, composed of name and status sub-elements.

cacheCluster_replicationGroupLogDeliveryEnabled :: Lens' CacheCluster (Maybe Bool)

A boolean value indicating whether log delivery is enabled for the replication group.

cacheCluster_authTokenEnabled :: Lens' CacheCluster (Maybe Bool)

A flag that enables using an AuthToken (password) when issuing Redis commands.

Default: false

cacheCluster_clientDownloadLandingPage :: Lens' CacheCluster (Maybe Text)

The URL of the web page where you can download the latest ElastiCache client library.

cacheCluster_preferredMaintenanceWindow :: Lens' CacheCluster (Maybe Text)

Specifies the weekly time range during which maintenance on the cluster is performed. It is specified as a range in the format ddd:hh24:mi-ddd:hh24:mi (24H Clock UTC). The minimum maintenance window is a 60 minute period.

Valid values for ddd are:

sun

mon

tue

wed

thu

fri

sat

Example: sun:23:00-mon:01:30

cacheCluster_cacheSubnetGroupName :: Lens' CacheCluster (Maybe Text)

The name of the cache subnet group associated with the cluster.

cacheCluster_preferredAvailabilityZone :: Lens' CacheCluster (Maybe Text)

The name of the Availability Zone in which the cluster is located or "Multiple" if the cache nodes are located in different Availability Zones.

cacheCluster_cacheParameterGroup :: Lens' CacheCluster (Maybe CacheParameterGroupStatus)

Status of the cache parameter group.

cacheCluster_cacheClusterStatus :: Lens' CacheCluster (Maybe Text)

The current state of this cluster, one of the following values: available, creating, deleted, deleting, incompatible-network, modifying, rebooting cluster nodes, restore-failed, or snapshotting.

cacheCluster_snapshotRetentionLimit :: Lens' CacheCluster (Maybe Int)

The number of days for which ElastiCache retains automatic cluster snapshots before deleting them. For example, if you set SnapshotRetentionLimit to 5, a snapshot that was taken today is retained for 5 days before being deleted.

If the value of SnapshotRetentionLimit is set to zero (0), backups are turned off.

cacheCluster_preferredOutpostArn :: Lens' CacheCluster (Maybe Text)

The outpost ARN in which the cache cluster is created.

cacheCluster_replicationGroupId :: Lens' CacheCluster (Maybe Text)

The replication group to which this cluster belongs. If this field is empty, the cluster is not associated with any replication group.

cacheCluster_pendingModifiedValues :: Lens' CacheCluster (Maybe PendingModifiedValues)

Undocumented member.

cacheCluster_numCacheNodes :: Lens' CacheCluster (Maybe Int)

The number of cache nodes in the cluster.

For clusters running Redis, this value must be 1. For clusters running Memcached, this value must be between 1 and 40.