Amazonka.ElastiCache.Types.CacheSecurityGroup

Description

Documentation

data CacheSecurityGroup

Represents the output of one of the following operations:

AuthorizeCacheSecurityGroupIngress

CreateCacheSecurityGroup

RevokeCacheSecurityGroupIngress

See: newCacheSecurityGroup smart constructor.

Constructors

CacheSecurityGroup'

Fields

cacheSecurityGroupName :: Maybe Text

The name of the cache security group.

arn :: Maybe Text

The ARN of the cache security group,

ownerId :: Maybe Text

The Amazon account ID of the cache security group owner.

eC2SecurityGroups :: Maybe [EC2SecurityGroup]

A list of Amazon EC2 security groups that are associated with this cache security group.

description :: Maybe Text

The description of the cache security group.

newCacheSecurityGroup :: CacheSecurityGroup

Create a value of CacheSecurityGroup with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:cacheSecurityGroupName:CacheSecurityGroup', cacheSecurityGroup_cacheSecurityGroupName - The name of the cache security group.

$sel:arn:CacheSecurityGroup', cacheSecurityGroup_arn - The ARN of the cache security group,

$sel:ownerId:CacheSecurityGroup', cacheSecurityGroup_ownerId - The Amazon account ID of the cache security group owner.

$sel:eC2SecurityGroups:CacheSecurityGroup', cacheSecurityGroup_eC2SecurityGroups - A list of Amazon EC2 security groups that are associated with this cache security group.

$sel:description:CacheSecurityGroup', cacheSecurityGroup_description - The description of the cache security group.

cacheSecurityGroup_cacheSecurityGroupName :: Lens' CacheSecurityGroup (Maybe Text)

The name of the cache security group.

cacheSecurityGroup_arn :: Lens' CacheSecurityGroup (Maybe Text)

The ARN of the cache security group,

cacheSecurityGroup_ownerId :: Lens' CacheSecurityGroup (Maybe Text)

The Amazon account ID of the cache security group owner.

cacheSecurityGroup_eC2SecurityGroups :: Lens' CacheSecurityGroup (Maybe [EC2SecurityGroup])

A list of Amazon EC2 security groups that are associated with this cache security group.

cacheSecurityGroup_description :: Lens' CacheSecurityGroup (Maybe Text)

The description of the cache security group.