Amazonka.ElastiCache.Types.GlobalReplicationGroup

Description

Documentation

data GlobalReplicationGroup

Consists of a primary cluster that accepts writes and an associated secondary cluster that resides in a different Amazon region. The secondary cluster accepts only reads. The primary cluster automatically replicates updates to the secondary cluster.

The GlobalReplicationGroupIdSuffix represents the name of the Global datastore, which is what you use to associate a secondary cluster.

See: newGlobalReplicationGroup smart constructor.

Constructors

GlobalReplicationGroup'

Fields

engineVersion :: Maybe Text

The Elasticache Redis engine version.

status :: Maybe Text

The status of the Global datastore

cacheNodeType :: Maybe Text

The cache node type of the Global datastore

clusterEnabled :: Maybe Bool

A flag that indicates whether the Global datastore is cluster enabled.

atRestEncryptionEnabled :: Maybe Bool

A flag that enables encryption at rest when set to true.

You cannot modify the value of AtRestEncryptionEnabled after the replication group is created. To enable encryption at rest on a replication group you must set AtRestEncryptionEnabled to true when you create the replication group.

Required: Only available when creating a replication group in an Amazon VPC using redis version 3.2.6, 4.x or later.

arn :: Maybe Text

The ARN (Amazon Resource Name) of the global replication group.

transitEncryptionEnabled :: Maybe Bool

A flag that enables in-transit encryption when set to true. You cannot modify the value of TransitEncryptionEnabled after the cluster is created. To enable in-transit encryption on a cluster you must set TransitEncryptionEnabled to true when you create a cluster.

Required: Only available when creating a replication group in an Amazon VPC using redis version 3.2.6, 4.x or later.

members :: Maybe [GlobalReplicationGroupMember]

The replication groups that comprise the Global datastore.

engine :: Maybe Text

The Elasticache engine. For Redis only.

authTokenEnabled :: Maybe Bool

A flag that enables using an AuthToken (password) when issuing Redis commands.

Default: false

globalNodeGroups :: Maybe [GlobalNodeGroup]

Indicates the slot configuration and global identifier for each slice group.

globalReplicationGroupId :: Maybe Text

The name of the Global datastore

globalReplicationGroupDescription :: Maybe Text

The optional description of the Global datastore

newGlobalReplicationGroup :: GlobalReplicationGroup

Create a value of GlobalReplicationGroup with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:engineVersion:GlobalReplicationGroup', globalReplicationGroup_engineVersion - The Elasticache Redis engine version.

$sel:status:GlobalReplicationGroup', globalReplicationGroup_status - The status of the Global datastore

$sel:cacheNodeType:GlobalReplicationGroup', globalReplicationGroup_cacheNodeType - The cache node type of the Global datastore

$sel:clusterEnabled:GlobalReplicationGroup', globalReplicationGroup_clusterEnabled - A flag that indicates whether the Global datastore is cluster enabled.

$sel:atRestEncryptionEnabled:GlobalReplicationGroup', globalReplicationGroup_atRestEncryptionEnabled - A flag that enables encryption at rest when set to true.

You cannot modify the value of AtRestEncryptionEnabled after the replication group is created. To enable encryption at rest on a replication group you must set AtRestEncryptionEnabled to true when you create the replication group.

Required: Only available when creating a replication group in an Amazon VPC using redis version 3.2.6, 4.x or later.

$sel:arn:GlobalReplicationGroup', globalReplicationGroup_arn - The ARN (Amazon Resource Name) of the global replication group.

$sel:transitEncryptionEnabled:GlobalReplicationGroup', globalReplicationGroup_transitEncryptionEnabled - A flag that enables in-transit encryption when set to true. You cannot modify the value of TransitEncryptionEnabled after the cluster is created. To enable in-transit encryption on a cluster you must set TransitEncryptionEnabled to true when you create a cluster.

Required: Only available when creating a replication group in an Amazon VPC using redis version 3.2.6, 4.x or later.

$sel:members:GlobalReplicationGroup', globalReplicationGroup_members - The replication groups that comprise the Global datastore.

$sel:engine:GlobalReplicationGroup', globalReplicationGroup_engine - The Elasticache engine. For Redis only.

$sel:authTokenEnabled:GlobalReplicationGroup', globalReplicationGroup_authTokenEnabled - A flag that enables using an AuthToken (password) when issuing Redis commands.

Default: false

$sel:globalNodeGroups:GlobalReplicationGroup', globalReplicationGroup_globalNodeGroups - Indicates the slot configuration and global identifier for each slice group.

$sel:globalReplicationGroupId:GlobalReplicationGroup', globalReplicationGroup_globalReplicationGroupId - The name of the Global datastore

$sel:globalReplicationGroupDescription:GlobalReplicationGroup', globalReplicationGroup_globalReplicationGroupDescription - The optional description of the Global datastore

globalReplicationGroup_engineVersion :: Lens' GlobalReplicationGroup (Maybe Text)

The Elasticache Redis engine version.

globalReplicationGroup_status :: Lens' GlobalReplicationGroup (Maybe Text)

The status of the Global datastore

globalReplicationGroup_cacheNodeType :: Lens' GlobalReplicationGroup (Maybe Text)

The cache node type of the Global datastore

globalReplicationGroup_clusterEnabled :: Lens' GlobalReplicationGroup (Maybe Bool)

A flag that indicates whether the Global datastore is cluster enabled.

globalReplicationGroup_atRestEncryptionEnabled :: Lens' GlobalReplicationGroup (Maybe Bool)

A flag that enables encryption at rest when set to true.

You cannot modify the value of AtRestEncryptionEnabled after the replication group is created. To enable encryption at rest on a replication group you must set AtRestEncryptionEnabled to true when you create the replication group.

Required: Only available when creating a replication group in an Amazon VPC using redis version 3.2.6, 4.x or later.

globalReplicationGroup_arn :: Lens' GlobalReplicationGroup (Maybe Text)

The ARN (Amazon Resource Name) of the global replication group.

globalReplicationGroup_transitEncryptionEnabled :: Lens' GlobalReplicationGroup (Maybe Bool)

A flag that enables in-transit encryption when set to true. You cannot modify the value of TransitEncryptionEnabled after the cluster is created. To enable in-transit encryption on a cluster you must set TransitEncryptionEnabled to true when you create a cluster.

Required: Only available when creating a replication group in an Amazon VPC using redis version 3.2.6, 4.x or later.

globalReplicationGroup_members :: Lens' GlobalReplicationGroup (Maybe [GlobalReplicationGroupMember])

The replication groups that comprise the Global datastore.

globalReplicationGroup_engine :: Lens' GlobalReplicationGroup (Maybe Text)

The Elasticache engine. For Redis only.

globalReplicationGroup_authTokenEnabled :: Lens' GlobalReplicationGroup (Maybe Bool)

A flag that enables using an AuthToken (password) when issuing Redis commands.

Default: false

globalReplicationGroup_globalNodeGroups :: Lens' GlobalReplicationGroup (Maybe [GlobalNodeGroup])

Indicates the slot configuration and global identifier for each slice group.

globalReplicationGroup_globalReplicationGroupId :: Lens' GlobalReplicationGroup (Maybe Text)

The name of the Global datastore

globalReplicationGroup_globalReplicationGroupDescription :: Lens' GlobalReplicationGroup (Maybe Text)

The optional description of the Global datastore