Amazonka.ElastiCache.Types.LogDeliveryConfigurationRequest

Description

Documentation

data LogDeliveryConfigurationRequest

Specifies the destination, format and type of the logs.

See: newLogDeliveryConfigurationRequest smart constructor.

Constructors

LogDeliveryConfigurationRequest'

Fields

logFormat :: Maybe LogFormat

Specifies either JSON or TEXT

enabled :: Maybe Bool

Specify if log delivery is enabled. Default true.

destinationDetails :: Maybe DestinationDetails

Configuration details of either a CloudWatch Logs destination or Kinesis Data Firehose destination.

logType :: Maybe LogType

Refers to slow-log.

destinationType :: Maybe DestinationType

Specify either cloudwatch-logs or kinesis-firehose as the destination type.

newLogDeliveryConfigurationRequest :: LogDeliveryConfigurationRequest

Create a value of LogDeliveryConfigurationRequest with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:logFormat:LogDeliveryConfigurationRequest', logDeliveryConfigurationRequest_logFormat - Specifies either JSON or TEXT

$sel:enabled:LogDeliveryConfigurationRequest', logDeliveryConfigurationRequest_enabled - Specify if log delivery is enabled. Default true.

$sel:destinationDetails:LogDeliveryConfigurationRequest', logDeliveryConfigurationRequest_destinationDetails - Configuration details of either a CloudWatch Logs destination or Kinesis Data Firehose destination.

$sel:logType:LogDeliveryConfigurationRequest', logDeliveryConfigurationRequest_logType - Refers to slow-log.

$sel:destinationType:LogDeliveryConfigurationRequest', logDeliveryConfigurationRequest_destinationType - Specify either cloudwatch-logs or kinesis-firehose as the destination type.

logDeliveryConfigurationRequest_logFormat :: Lens' LogDeliveryConfigurationRequest (Maybe LogFormat)

Specifies either JSON or TEXT

logDeliveryConfigurationRequest_enabled :: Lens' LogDeliveryConfigurationRequest (Maybe Bool)

Specify if log delivery is enabled. Default true.

logDeliveryConfigurationRequest_destinationDetails :: Lens' LogDeliveryConfigurationRequest (Maybe DestinationDetails)

Configuration details of either a CloudWatch Logs destination or Kinesis Data Firehose destination.

logDeliveryConfigurationRequest_logType :: Lens' LogDeliveryConfigurationRequest (Maybe LogType)

Refers to slow-log.

logDeliveryConfigurationRequest_destinationType :: Lens' LogDeliveryConfigurationRequest (Maybe DestinationType)

Specify either cloudwatch-logs or kinesis-firehose as the destination type.