Amazonka.ElastiCache.Types.NodeGroupConfiguration

Description

Documentation

data NodeGroupConfiguration

Node group (shard) configuration options. Each node group (shard) configuration has the following: Slots, PrimaryAvailabilityZone, ReplicaAvailabilityZones, ReplicaCount.

See: newNodeGroupConfiguration smart constructor.

Constructors

NodeGroupConfiguration'

Fields

slots :: Maybe Text

A string that specifies the keyspace for a particular node group. Keyspaces range from 0 to 16,383. The string is in the format startkey-endkey.

Example: "0-3999"

replicaOutpostArns :: Maybe [Text]

The outpost ARN of the node replicas.

replicaCount :: Maybe Int

The number of read replica nodes in this node group (shard).

primaryAvailabilityZone :: Maybe Text

The Availability Zone where the primary node of this node group (shard) is launched.

replicaAvailabilityZones :: Maybe [Text]

A list of Availability Zones to be used for the read replicas. The number of Availability Zones in this list must match the value of ReplicaCount or ReplicasPerNodeGroup if not specified.

primaryOutpostArn :: Maybe Text

The outpost ARN of the primary node.

nodeGroupId :: Maybe Text

Either the ElastiCache for Redis supplied 4-digit id or a user supplied id for the node group these configuration values apply to.

newNodeGroupConfiguration :: NodeGroupConfiguration

Create a value of NodeGroupConfiguration with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:slots:NodeGroupConfiguration', nodeGroupConfiguration_slots - A string that specifies the keyspace for a particular node group. Keyspaces range from 0 to 16,383. The string is in the format startkey-endkey.

Example: "0-3999"

$sel:replicaOutpostArns:NodeGroupConfiguration', nodeGroupConfiguration_replicaOutpostArns - The outpost ARN of the node replicas.

$sel:replicaCount:NodeGroupConfiguration', nodeGroupConfiguration_replicaCount - The number of read replica nodes in this node group (shard).

$sel:primaryAvailabilityZone:NodeGroupConfiguration', nodeGroupConfiguration_primaryAvailabilityZone - The Availability Zone where the primary node of this node group (shard) is launched.

$sel:replicaAvailabilityZones:NodeGroupConfiguration', nodeGroupConfiguration_replicaAvailabilityZones - A list of Availability Zones to be used for the read replicas. The number of Availability Zones in this list must match the value of ReplicaCount or ReplicasPerNodeGroup if not specified.

$sel:primaryOutpostArn:NodeGroupConfiguration', nodeGroupConfiguration_primaryOutpostArn - The outpost ARN of the primary node.

$sel:nodeGroupId:NodeGroupConfiguration', nodeGroupConfiguration_nodeGroupId - Either the ElastiCache for Redis supplied 4-digit id or a user supplied id for the node group these configuration values apply to.

nodeGroupConfiguration_slots :: Lens' NodeGroupConfiguration (Maybe Text)

A string that specifies the keyspace for a particular node group. Keyspaces range from 0 to 16,383. The string is in the format startkey-endkey.

Example: "0-3999"

nodeGroupConfiguration_replicaOutpostArns :: Lens' NodeGroupConfiguration (Maybe [Text])

The outpost ARN of the node replicas.

nodeGroupConfiguration_replicaCount :: Lens' NodeGroupConfiguration (Maybe Int)

The number of read replica nodes in this node group (shard).

nodeGroupConfiguration_primaryAvailabilityZone :: Lens' NodeGroupConfiguration (Maybe Text)

The Availability Zone where the primary node of this node group (shard) is launched.

nodeGroupConfiguration_replicaAvailabilityZones :: Lens' NodeGroupConfiguration (Maybe [Text])

A list of Availability Zones to be used for the read replicas. The number of Availability Zones in this list must match the value of ReplicaCount or ReplicasPerNodeGroup if not specified.

nodeGroupConfiguration_primaryOutpostArn :: Lens' NodeGroupConfiguration (Maybe Text)

The outpost ARN of the primary node.

nodeGroupConfiguration_nodeGroupId :: Lens' NodeGroupConfiguration (Maybe Text)

Either the ElastiCache for Redis supplied 4-digit id or a user supplied id for the node group these configuration values apply to.