Amazonka.ElasticBeanstalk.CreatePlatformVersion

Description

Create a new version of your custom platform.

Creating a Request

data CreatePlatformVersion

Request to create a new platform version.

See: newCreatePlatformVersion smart constructor.

Constructors

CreatePlatformVersion'

Fields

optionSettings :: Maybe [ConfigurationOptionSetting]

The configuration option settings to apply to the builder environment.

environmentName :: Maybe Text

The name of the builder environment.

tags :: Maybe [Tag]

Specifies the tags applied to the new platform version.

Elastic Beanstalk applies these tags only to the platform version. Environments that you create using the platform version don't inherit the tags.

platformName :: Text

The name of your custom platform.

platformVersion :: Text

The number, such as 1.0.2, for the new platform version.

platformDefinitionBundle :: S3Location

The location of the platform definition archive in Amazon S3.

newCreatePlatformVersion

Arguments

:: Text
$sel:platformName:CreatePlatformVersion'

-> Text
$sel:platformVersion:CreatePlatformVersion'

-> S3Location
$sel:platformDefinitionBundle:CreatePlatformVersion'

-> CreatePlatformVersion

Create a value of CreatePlatformVersion with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:optionSettings:CreatePlatformVersion', createPlatformVersion_optionSettings - The configuration option settings to apply to the builder environment.

$sel:environmentName:CreatePlatformVersion', createPlatformVersion_environmentName - The name of the builder environment.

$sel:tags:CreatePlatformVersion', createPlatformVersion_tags - Specifies the tags applied to the new platform version.

Elastic Beanstalk applies these tags only to the platform version. Environments that you create using the platform version don't inherit the tags.

$sel:platformName:CreatePlatformVersion', createPlatformVersion_platformName - The name of your custom platform.

$sel:platformVersion:CreatePlatformVersion', createPlatformVersion_platformVersion - The number, such as 1.0.2, for the new platform version.

$sel:platformDefinitionBundle:CreatePlatformVersion', createPlatformVersion_platformDefinitionBundle - The location of the platform definition archive in Amazon S3.

Request Lenses

createPlatformVersion_optionSettings :: Lens' CreatePlatformVersion (Maybe [ConfigurationOptionSetting])

The configuration option settings to apply to the builder environment.

createPlatformVersion_environmentName :: Lens' CreatePlatformVersion (Maybe Text)

The name of the builder environment.

createPlatformVersion_tags :: Lens' CreatePlatformVersion (Maybe [Tag])

Specifies the tags applied to the new platform version.

Elastic Beanstalk applies these tags only to the platform version. Environments that you create using the platform version don't inherit the tags.

createPlatformVersion_platformName :: Lens' CreatePlatformVersion Text

The name of your custom platform.

createPlatformVersion_platformVersion :: Lens' CreatePlatformVersion Text

The number, such as 1.0.2, for the new platform version.

createPlatformVersion_platformDefinitionBundle :: Lens' CreatePlatformVersion S3Location

The location of the platform definition archive in Amazon S3.

Destructuring the Response

data CreatePlatformVersionResponse

See: newCreatePlatformVersionResponse smart constructor.

Constructors

CreatePlatformVersionResponse'

Fields

builder :: Maybe Builder

The builder used to create the custom platform.

platformSummary :: Maybe PlatformSummary

Detailed information about the new version of the custom platform.

httpStatus :: Int

The response's http status code.

newCreatePlatformVersionResponse

Arguments

:: Int
$sel:httpStatus:CreatePlatformVersionResponse'

-> CreatePlatformVersionResponse

Create a value of CreatePlatformVersionResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:builder:CreatePlatformVersionResponse', createPlatformVersionResponse_builder - The builder used to create the custom platform.

$sel:platformSummary:CreatePlatformVersionResponse', createPlatformVersionResponse_platformSummary - Detailed information about the new version of the custom platform.

$sel:httpStatus:CreatePlatformVersionResponse', createPlatformVersionResponse_httpStatus - The response's http status code.

Response Lenses

createPlatformVersionResponse_builder :: Lens' CreatePlatformVersionResponse (Maybe Builder)

The builder used to create the custom platform.

createPlatformVersionResponse_platformSummary :: Lens' CreatePlatformVersionResponse (Maybe PlatformSummary)

Detailed information about the new version of the custom platform.

createPlatformVersionResponse_httpStatus :: Lens' CreatePlatformVersionResponse Int

The response's http status code.