Amazonka.ElasticBeanstalk.DescribeEnvironments

Description

Returns descriptions for existing environments.

This operation returns paginated results.

Creating a Request

data DescribeEnvironments

Request to describe one or more environments.

See: newDescribeEnvironments smart constructor.

Constructors

DescribeEnvironments'

Fields

environmentIds :: Maybe [Text]

If specified, AWS Elastic Beanstalk restricts the returned descriptions to include only those that have the specified IDs.

environmentNames :: Maybe [Text]

If specified, AWS Elastic Beanstalk restricts the returned descriptions to include only those that have the specified names.

nextToken :: Maybe Text

For a paginated request. Specify a token from a previous response page to retrieve the next response page. All other parameter values must be identical to the ones specified in the initial request.

If no NextToken is specified, the first page is retrieved.

versionLabel :: Maybe Text

If specified, AWS Elastic Beanstalk restricts the returned descriptions to include only those that are associated with this application version.

maxRecords :: Maybe Natural

For a paginated request. Specify a maximum number of environments to include in each response.

If no MaxRecords is specified, all available environments are retrieved in a single response.

applicationName :: Maybe Text

If specified, AWS Elastic Beanstalk restricts the returned descriptions to include only those that are associated with this application.

includedDeletedBackTo :: Maybe ISO8601

If specified when IncludeDeleted is set to true, then environments deleted after this date are displayed.

includeDeleted :: Maybe Bool

Indicates whether to include deleted environments:

true: Environments that have been deleted after IncludedDeletedBackTo are displayed.

false: Do not include deleted environments.

newDescribeEnvironments :: DescribeEnvironments

Create a value of DescribeEnvironments with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:environmentIds:DescribeEnvironments', describeEnvironments_environmentIds - If specified, AWS Elastic Beanstalk restricts the returned descriptions to include only those that have the specified IDs.

$sel:environmentNames:DescribeEnvironments', describeEnvironments_environmentNames - If specified, AWS Elastic Beanstalk restricts the returned descriptions to include only those that have the specified names.

$sel:nextToken:DescribeEnvironments', describeEnvironments_nextToken - For a paginated request. Specify a token from a previous response page to retrieve the next response page. All other parameter values must be identical to the ones specified in the initial request.

If no NextToken is specified, the first page is retrieved.

$sel:versionLabel:DescribeEnvironments', describeEnvironments_versionLabel - If specified, AWS Elastic Beanstalk restricts the returned descriptions to include only those that are associated with this application version.

$sel:maxRecords:DescribeEnvironments', describeEnvironments_maxRecords - For a paginated request. Specify a maximum number of environments to include in each response.

If no MaxRecords is specified, all available environments are retrieved in a single response.

$sel:applicationName:DescribeEnvironments', describeEnvironments_applicationName - If specified, AWS Elastic Beanstalk restricts the returned descriptions to include only those that are associated with this application.

$sel:includedDeletedBackTo:DescribeEnvironments', describeEnvironments_includedDeletedBackTo - If specified when IncludeDeleted is set to true, then environments deleted after this date are displayed.

$sel:includeDeleted:DescribeEnvironments', describeEnvironments_includeDeleted - Indicates whether to include deleted environments:

true: Environments that have been deleted after IncludedDeletedBackTo are displayed.

false: Do not include deleted environments.

Request Lenses

describeEnvironments_environmentIds :: Lens' DescribeEnvironments (Maybe [Text])

If specified, AWS Elastic Beanstalk restricts the returned descriptions to include only those that have the specified IDs.

describeEnvironments_environmentNames :: Lens' DescribeEnvironments (Maybe [Text])

If specified, AWS Elastic Beanstalk restricts the returned descriptions to include only those that have the specified names.

describeEnvironments_nextToken :: Lens' DescribeEnvironments (Maybe Text)

For a paginated request. Specify a token from a previous response page to retrieve the next response page. All other parameter values must be identical to the ones specified in the initial request.

If no NextToken is specified, the first page is retrieved.

describeEnvironments_versionLabel :: Lens' DescribeEnvironments (Maybe Text)

If specified, AWS Elastic Beanstalk restricts the returned descriptions to include only those that are associated with this application version.

describeEnvironments_maxRecords :: Lens' DescribeEnvironments (Maybe Natural)

For a paginated request. Specify a maximum number of environments to include in each response.

If no MaxRecords is specified, all available environments are retrieved in a single response.

describeEnvironments_applicationName :: Lens' DescribeEnvironments (Maybe Text)

If specified, AWS Elastic Beanstalk restricts the returned descriptions to include only those that are associated with this application.

describeEnvironments_includedDeletedBackTo :: Lens' DescribeEnvironments (Maybe UTCTime)

If specified when IncludeDeleted is set to true, then environments deleted after this date are displayed.

describeEnvironments_includeDeleted :: Lens' DescribeEnvironments (Maybe Bool)

Indicates whether to include deleted environments:

true: Environments that have been deleted after IncludedDeletedBackTo are displayed.

false: Do not include deleted environments.

Destructuring the Response

data EnvironmentDescriptionsMessage

Result message containing a list of environment descriptions.

See: newEnvironmentDescriptionsMessage smart constructor.

Constructors

EnvironmentDescriptionsMessage'

Fields

nextToken :: Maybe Text

In a paginated request, the token that you can pass in a subsequent request to get the next response page.

environments :: Maybe [EnvironmentDescription]

Returns an EnvironmentDescription list.

newEnvironmentDescriptionsMessage :: EnvironmentDescriptionsMessage

Create a value of EnvironmentDescriptionsMessage with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:nextToken:EnvironmentDescriptionsMessage', environmentDescriptionsMessage_nextToken - In a paginated request, the token that you can pass in a subsequent request to get the next response page.

$sel:environments:EnvironmentDescriptionsMessage', environmentDescriptionsMessage_environments - Returns an EnvironmentDescription list.

Response Lenses

environmentDescriptionsMessage_nextToken :: Lens' EnvironmentDescriptionsMessage (Maybe Text)

In a paginated request, the token that you can pass in a subsequent request to get the next response page.

environmentDescriptionsMessage_environments :: Lens' EnvironmentDescriptionsMessage (Maybe [EnvironmentDescription])

Returns an EnvironmentDescription list.