Amazonka.ElasticBeanstalk.ListPlatformBranches

Description

Lists the platform branches available for your account in an AWS Region. Provides summary information about each platform branch.

For definitions of platform branch and other platform-related terms, see AWS Elastic Beanstalk Platforms Glossary.

Creating a Request

data ListPlatformBranches

See: newListPlatformBranches smart constructor.

Constructors

ListPlatformBranches'

Fields

filters :: Maybe [SearchFilter]

Criteria for restricting the resulting list of platform branches. The filter is evaluated as a logical conjunction (AND) of the separate SearchFilter terms.

The following list shows valid attribute values for each of the SearchFilter terms. Most operators take a single value. The in and not_in operators can take multiple values.

Attribute = BranchName:

Operator: = | != | begins_with | ends_with | contains | in | not_in

Attribute = LifecycleState:

Operator: = | != | in | not_in

Values: beta | supported | deprecated | retired

Attribute = PlatformName:

Operator: = | != | begins_with | ends_with | contains | in | not_in

Attribute = TierType:

Operator: = | !=

Values: WebServer/Standard | Worker/SQS/HTTP

Array size: limited to 10 SearchFilter objects.

Within each SearchFilter item, the Values array is limited to 10 items.

nextToken :: Maybe Text

For a paginated request. Specify a token from a previous response page to retrieve the next response page. All other parameter values must be identical to the ones specified in the initial request.

If no NextToken is specified, the first page is retrieved.

maxRecords :: Maybe Natural

The maximum number of platform branch values returned in one call.

newListPlatformBranches :: ListPlatformBranches

Create a value of ListPlatformBranches with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:filters:ListPlatformBranches', listPlatformBranches_filters - Criteria for restricting the resulting list of platform branches. The filter is evaluated as a logical conjunction (AND) of the separate SearchFilter terms.

The following list shows valid attribute values for each of the SearchFilter terms. Most operators take a single value. The in and not_in operators can take multiple values.

Attribute = BranchName:

Operator: = | != | begins_with | ends_with | contains | in | not_in

Attribute = LifecycleState:

Operator: = | != | in | not_in

Values: beta | supported | deprecated | retired

Attribute = PlatformName:

Operator: = | != | begins_with | ends_with | contains | in | not_in

Attribute = TierType:

Operator: = | !=

Values: WebServer/Standard | Worker/SQS/HTTP

Array size: limited to 10 SearchFilter objects.

Within each SearchFilter item, the Values array is limited to 10 items.

$sel:nextToken:ListPlatformBranches', listPlatformBranches_nextToken - For a paginated request. Specify a token from a previous response page to retrieve the next response page. All other parameter values must be identical to the ones specified in the initial request.

If no NextToken is specified, the first page is retrieved.

$sel:maxRecords:ListPlatformBranches', listPlatformBranches_maxRecords - The maximum number of platform branch values returned in one call.

Request Lenses

listPlatformBranches_filters :: Lens' ListPlatformBranches (Maybe [SearchFilter])

Criteria for restricting the resulting list of platform branches. The filter is evaluated as a logical conjunction (AND) of the separate SearchFilter terms.

The following list shows valid attribute values for each of the SearchFilter terms. Most operators take a single value. The in and not_in operators can take multiple values.

Attribute = BranchName:

Operator: = | != | begins_with | ends_with | contains | in | not_in

Attribute = LifecycleState:

Operator: = | != | in | not_in

Values: beta | supported | deprecated | retired

Attribute = PlatformName:

Operator: = | != | begins_with | ends_with | contains | in | not_in

Attribute = TierType:

Operator: = | !=

Values: WebServer/Standard | Worker/SQS/HTTP

Array size: limited to 10 SearchFilter objects.

Within each SearchFilter item, the Values array is limited to 10 items.

listPlatformBranches_nextToken :: Lens' ListPlatformBranches (Maybe Text)

For a paginated request. Specify a token from a previous response page to retrieve the next response page. All other parameter values must be identical to the ones specified in the initial request.

If no NextToken is specified, the first page is retrieved.

listPlatformBranches_maxRecords :: Lens' ListPlatformBranches (Maybe Natural)

The maximum number of platform branch values returned in one call.

Destructuring the Response

data ListPlatformBranchesResponse

See: newListPlatformBranchesResponse smart constructor.

Constructors

ListPlatformBranchesResponse'

Fields

platformBranchSummaryList :: Maybe [PlatformBranchSummary]

Summary information about the platform branches.

nextToken :: Maybe Text

In a paginated request, if this value isn't null, it's the token that you can pass in a subsequent request to get the next response page.

httpStatus :: Int

The response's http status code.

newListPlatformBranchesResponse

Arguments

:: Int
$sel:httpStatus:ListPlatformBranchesResponse'

-> ListPlatformBranchesResponse

Create a value of ListPlatformBranchesResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:platformBranchSummaryList:ListPlatformBranchesResponse', listPlatformBranchesResponse_platformBranchSummaryList - Summary information about the platform branches.

$sel:nextToken:ListPlatformBranches', listPlatformBranchesResponse_nextToken - In a paginated request, if this value isn't null, it's the token that you can pass in a subsequent request to get the next response page.

$sel:httpStatus:ListPlatformBranchesResponse', listPlatformBranchesResponse_httpStatus - The response's http status code.

Response Lenses

listPlatformBranchesResponse_platformBranchSummaryList :: Lens' ListPlatformBranchesResponse (Maybe [PlatformBranchSummary])

Summary information about the platform branches.

listPlatformBranchesResponse_nextToken :: Lens' ListPlatformBranchesResponse (Maybe Text)

In a paginated request, if this value isn't null, it's the token that you can pass in a subsequent request to get the next response page.

listPlatformBranchesResponse_httpStatus :: Lens' ListPlatformBranchesResponse Int

The response's http status code.