Amazonka.ElasticBeanstalk.RequestEnvironmentInfo

Description

Initiates a request to compile the specified type of information of the deployed environment.

Setting the InfoType to tail compiles the last lines from the application server log files of every Amazon EC2 instance in your environment.

Setting the InfoType to bundle compresses the application server log files for every Amazon EC2 instance into a .zip file. Legacy and .NET containers do not support bundle logs.

Use RetrieveEnvironmentInfo to obtain the set of logs.

Related Topics

RetrieveEnvironmentInfo

Creating a Request

data RequestEnvironmentInfo

Request to retrieve logs from an environment and store them in your Elastic Beanstalk storage bucket.

See: newRequestEnvironmentInfo smart constructor.

Constructors

RequestEnvironmentInfo'

Fields

environmentName :: Maybe Text

The name of the environment of the requested data.

If no such environment is found, RequestEnvironmentInfo returns an InvalidParameterValue error.

Condition: You must specify either this or an EnvironmentId, or both. If you do not specify either, AWS Elastic Beanstalk returns MissingRequiredParameter error.

environmentId :: Maybe Text

The ID of the environment of the requested data.

If no such environment is found, RequestEnvironmentInfo returns an InvalidParameterValue error.

Condition: You must specify either this or an EnvironmentName, or both. If you do not specify either, AWS Elastic Beanstalk returns MissingRequiredParameter error.

infoType :: EnvironmentInfoType

The type of information to request.

newRequestEnvironmentInfo

Arguments

:: EnvironmentInfoType
$sel:infoType:RequestEnvironmentInfo'

-> RequestEnvironmentInfo

Create a value of RequestEnvironmentInfo with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:environmentName:RequestEnvironmentInfo', requestEnvironmentInfo_environmentName - The name of the environment of the requested data.

If no such environment is found, RequestEnvironmentInfo returns an InvalidParameterValue error.

Condition: You must specify either this or an EnvironmentId, or both. If you do not specify either, AWS Elastic Beanstalk returns MissingRequiredParameter error.

$sel:environmentId:RequestEnvironmentInfo', requestEnvironmentInfo_environmentId - The ID of the environment of the requested data.

If no such environment is found, RequestEnvironmentInfo returns an InvalidParameterValue error.

Condition: You must specify either this or an EnvironmentName, or both. If you do not specify either, AWS Elastic Beanstalk returns MissingRequiredParameter error.

$sel:infoType:RequestEnvironmentInfo', requestEnvironmentInfo_infoType - The type of information to request.

Request Lenses

requestEnvironmentInfo_environmentName :: Lens' RequestEnvironmentInfo (Maybe Text)

The name of the environment of the requested data.

If no such environment is found, RequestEnvironmentInfo returns an InvalidParameterValue error.

Condition: You must specify either this or an EnvironmentId, or both. If you do not specify either, AWS Elastic Beanstalk returns MissingRequiredParameter error.

requestEnvironmentInfo_environmentId :: Lens' RequestEnvironmentInfo (Maybe Text)

The ID of the environment of the requested data.

If no such environment is found, RequestEnvironmentInfo returns an InvalidParameterValue error.

Condition: You must specify either this or an EnvironmentName, or both. If you do not specify either, AWS Elastic Beanstalk returns MissingRequiredParameter error.

requestEnvironmentInfo_infoType :: Lens' RequestEnvironmentInfo EnvironmentInfoType

The type of information to request.

Destructuring the Response

data RequestEnvironmentInfoResponse

See: newRequestEnvironmentInfoResponse smart constructor.

Constructors

RequestEnvironmentInfoResponse'

newRequestEnvironmentInfoResponse :: RequestEnvironmentInfoResponse

Create a value of RequestEnvironmentInfoResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.