Amazonka.ElasticBeanstalk.Types.ApplicationMetrics

Description

Documentation

data ApplicationMetrics

Application request metrics for an AWS Elastic Beanstalk environment.

See: newApplicationMetrics smart constructor.

Constructors

ApplicationMetrics'

Fields

requestCount :: Maybe Int

Average number of requests handled by the web server per second over the last 10 seconds.

latency :: Maybe Latency

Represents the average latency for the slowest X percent of requests over the last 10 seconds. Latencies are in seconds with one millisecond resolution.

statusCodes :: Maybe StatusCodes

Represents the percentage of requests over the last 10 seconds that resulted in each type of status code response.

duration :: Maybe Int

The amount of time that the metrics cover (usually 10 seconds). For example, you might have 5 requests (request_count) within the most recent time slice of 10 seconds (duration).

newApplicationMetrics :: ApplicationMetrics

Create a value of ApplicationMetrics with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:requestCount:ApplicationMetrics', applicationMetrics_requestCount - Average number of requests handled by the web server per second over the last 10 seconds.

$sel:latency:ApplicationMetrics', applicationMetrics_latency - Represents the average latency for the slowest X percent of requests over the last 10 seconds. Latencies are in seconds with one millisecond resolution.

$sel:statusCodes:ApplicationMetrics', applicationMetrics_statusCodes - Represents the percentage of requests over the last 10 seconds that resulted in each type of status code response.

$sel:duration:ApplicationMetrics', applicationMetrics_duration - The amount of time that the metrics cover (usually 10 seconds). For example, you might have 5 requests (request_count) within the most recent time slice of 10 seconds (duration).

applicationMetrics_requestCount :: Lens' ApplicationMetrics (Maybe Int)

Average number of requests handled by the web server per second over the last 10 seconds.

applicationMetrics_latency :: Lens' ApplicationMetrics (Maybe Latency)

Represents the average latency for the slowest X percent of requests over the last 10 seconds. Latencies are in seconds with one millisecond resolution.

applicationMetrics_statusCodes :: Lens' ApplicationMetrics (Maybe StatusCodes)

Represents the percentage of requests over the last 10 seconds that resulted in each type of status code response.

applicationMetrics_duration :: Lens' ApplicationMetrics (Maybe Int)

The amount of time that the metrics cover (usually 10 seconds). For example, you might have 5 requests (request_count) within the most recent time slice of 10 seconds (duration).