Amazonka.ElasticBeanstalk.Types.ApplicationVersionDescription

Description

Documentation

data ApplicationVersionDescription

Describes the properties of an application version.

See: newApplicationVersionDescription smart constructor.

Constructors

ApplicationVersionDescription'

Fields

status :: Maybe ApplicationVersionStatus

The processing status of the application version. Reflects the state of the application version during its creation. Many of the values are only applicable if you specified True for the Process parameter of the CreateApplicationVersion action. The following list describes the possible values.

Unprocessed – Application version wasn't pre-processed or validated. Elastic Beanstalk will validate configuration files during deployment of the application version to an environment.

Processing – Elastic Beanstalk is currently processing the application version.

Building – Application version is currently undergoing an AWS CodeBuild build.

Processed – Elastic Beanstalk was successfully pre-processed and validated.

Failed – Either the AWS CodeBuild build failed or configuration files didn't pass validation. This application version isn't usable.

sourceBundle :: Maybe S3Location

The storage location of the application version's source bundle in Amazon S3.

dateUpdated :: Maybe ISO8601

The last modified date of the application version.

dateCreated :: Maybe ISO8601

The creation date of the application version.

versionLabel :: Maybe Text

A unique identifier for the application version.

sourceBuildInformation :: Maybe SourceBuildInformation

If the version's source code was retrieved from AWS CodeCommit, the location of the source code for the application version.

applicationName :: Maybe Text

The name of the application to which the application version belongs.

applicationVersionArn :: Maybe Text

The Amazon Resource Name (ARN) of the application version.

buildArn :: Maybe Text

Reference to the artifact from the AWS CodeBuild build.

description :: Maybe Text

The description of the application version.

newApplicationVersionDescription :: ApplicationVersionDescription

Create a value of ApplicationVersionDescription with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:status:ApplicationVersionDescription', applicationVersionDescription_status - The processing status of the application version. Reflects the state of the application version during its creation. Many of the values are only applicable if you specified True for the Process parameter of the CreateApplicationVersion action. The following list describes the possible values.

Unprocessed – Application version wasn't pre-processed or validated. Elastic Beanstalk will validate configuration files during deployment of the application version to an environment.

Processing – Elastic Beanstalk is currently processing the application version.

Building – Application version is currently undergoing an AWS CodeBuild build.

Processed – Elastic Beanstalk was successfully pre-processed and validated.

Failed – Either the AWS CodeBuild build failed or configuration files didn't pass validation. This application version isn't usable.

$sel:sourceBundle:ApplicationVersionDescription', applicationVersionDescription_sourceBundle - The storage location of the application version's source bundle in Amazon S3.

$sel:dateUpdated:ApplicationVersionDescription', applicationVersionDescription_dateUpdated - The last modified date of the application version.

$sel:dateCreated:ApplicationVersionDescription', applicationVersionDescription_dateCreated - The creation date of the application version.

$sel:versionLabel:ApplicationVersionDescription', applicationVersionDescription_versionLabel - A unique identifier for the application version.

$sel:sourceBuildInformation:ApplicationVersionDescription', applicationVersionDescription_sourceBuildInformation - If the version's source code was retrieved from AWS CodeCommit, the location of the source code for the application version.

$sel:applicationName:ApplicationVersionDescription', applicationVersionDescription_applicationName - The name of the application to which the application version belongs.

$sel:applicationVersionArn:ApplicationVersionDescription', applicationVersionDescription_applicationVersionArn - The Amazon Resource Name (ARN) of the application version.

$sel:buildArn:ApplicationVersionDescription', applicationVersionDescription_buildArn - Reference to the artifact from the AWS CodeBuild build.

$sel:description:ApplicationVersionDescription', applicationVersionDescription_description - The description of the application version.

applicationVersionDescription_status :: Lens' ApplicationVersionDescription (Maybe ApplicationVersionStatus)

The processing status of the application version. Reflects the state of the application version during its creation. Many of the values are only applicable if you specified True for the Process parameter of the CreateApplicationVersion action. The following list describes the possible values.

Unprocessed – Application version wasn't pre-processed or validated. Elastic Beanstalk will validate configuration files during deployment of the application version to an environment.

Processing – Elastic Beanstalk is currently processing the application version.

Building – Application version is currently undergoing an AWS CodeBuild build.

Processed – Elastic Beanstalk was successfully pre-processed and validated.

Failed – Either the AWS CodeBuild build failed or configuration files didn't pass validation. This application version isn't usable.

applicationVersionDescription_sourceBundle :: Lens' ApplicationVersionDescription (Maybe S3Location)

The storage location of the application version's source bundle in Amazon S3.

applicationVersionDescription_dateUpdated :: Lens' ApplicationVersionDescription (Maybe UTCTime)

The last modified date of the application version.

applicationVersionDescription_dateCreated :: Lens' ApplicationVersionDescription (Maybe UTCTime)

The creation date of the application version.

applicationVersionDescription_versionLabel :: Lens' ApplicationVersionDescription (Maybe Text)

A unique identifier for the application version.

applicationVersionDescription_sourceBuildInformation :: Lens' ApplicationVersionDescription (Maybe SourceBuildInformation)

If the version's source code was retrieved from AWS CodeCommit, the location of the source code for the application version.

applicationVersionDescription_applicationName :: Lens' ApplicationVersionDescription (Maybe Text)

The name of the application to which the application version belongs.

applicationVersionDescription_applicationVersionArn :: Lens' ApplicationVersionDescription (Maybe Text)

The Amazon Resource Name (ARN) of the application version.

applicationVersionDescription_buildArn :: Lens' ApplicationVersionDescription (Maybe Text)

Reference to the artifact from the AWS CodeBuild build.

applicationVersionDescription_description :: Lens' ApplicationVersionDescription (Maybe Text)

The description of the application version.