Amazonka.ElasticBeanstalk.Types.PlatformDescription

Description

Documentation

data PlatformDescription

Detailed information about a platform version.

See: newPlatformDescription smart constructor.

Constructors

PlatformDescription'

Fields

platformBranchName :: Maybe Text

The platform branch to which the platform version belongs.

supportedAddonList :: Maybe [Text]

The additions supported by the platform version.

platformCategory :: Maybe Text

The category of the platform version.

platformBranchLifecycleState :: Maybe Text

The state of the platform version's branch in its lifecycle.

Possible values: Beta | Supported | Deprecated | Retired

platformVersion :: Maybe Text

The version of the platform version.

platformStatus :: Maybe PlatformStatus

The status of the platform version.

maintainer :: Maybe Text

Information about the maintainer of the platform version.

platformLifecycleState :: Maybe Text

The state of the platform version in its lifecycle.

Possible values: Recommended | null

If a null value is returned, the platform version isn't the recommended one for its branch. Each platform branch has a single recommended platform version, typically the most recent one.

platformOwner :: Maybe Text

The AWS account ID of the person who created the platform version.

dateUpdated :: Maybe ISO8601

The date when the platform version was last updated.

customAmiList :: Maybe [CustomAmi]

The custom AMIs supported by the platform version.

dateCreated :: Maybe ISO8601

The date when the platform version was created.

operatingSystemName :: Maybe Text

The operating system used by the platform version.

frameworks :: Maybe [PlatformFramework]

The frameworks supported by the platform version.

platformArn :: Maybe Text

The ARN of the platform version.

operatingSystemVersion :: Maybe Text

The version of the operating system used by the platform version.

programmingLanguages :: Maybe [PlatformProgrammingLanguage]

The programming languages supported by the platform version.

solutionStackName :: Maybe Text

The name of the solution stack used by the platform version.

platformName :: Maybe Text

The name of the platform version.

description :: Maybe Text

The description of the platform version.

supportedTierList :: Maybe [Text]

The tiers supported by the platform version.

newPlatformDescription :: PlatformDescription

Create a value of PlatformDescription with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:platformBranchName:PlatformDescription', platformDescription_platformBranchName - The platform branch to which the platform version belongs.

$sel:supportedAddonList:PlatformDescription', platformDescription_supportedAddonList - The additions supported by the platform version.

$sel:platformCategory:PlatformDescription', platformDescription_platformCategory - The category of the platform version.

$sel:platformBranchLifecycleState:PlatformDescription', platformDescription_platformBranchLifecycleState - The state of the platform version's branch in its lifecycle.

Possible values: Beta | Supported | Deprecated | Retired

$sel:platformVersion:PlatformDescription', platformDescription_platformVersion - The version of the platform version.

$sel:platformStatus:PlatformDescription', platformDescription_platformStatus - The status of the platform version.

$sel:maintainer:PlatformDescription', platformDescription_maintainer - Information about the maintainer of the platform version.

$sel:platformLifecycleState:PlatformDescription', platformDescription_platformLifecycleState - The state of the platform version in its lifecycle.

Possible values: Recommended | null

If a null value is returned, the platform version isn't the recommended one for its branch. Each platform branch has a single recommended platform version, typically the most recent one.

$sel:platformOwner:PlatformDescription', platformDescription_platformOwner - The AWS account ID of the person who created the platform version.

$sel:dateUpdated:PlatformDescription', platformDescription_dateUpdated - The date when the platform version was last updated.

$sel:customAmiList:PlatformDescription', platformDescription_customAmiList - The custom AMIs supported by the platform version.

$sel:dateCreated:PlatformDescription', platformDescription_dateCreated - The date when the platform version was created.

$sel:operatingSystemName:PlatformDescription', platformDescription_operatingSystemName - The operating system used by the platform version.

$sel:frameworks:PlatformDescription', platformDescription_frameworks - The frameworks supported by the platform version.

$sel:platformArn:PlatformDescription', platformDescription_platformArn - The ARN of the platform version.

$sel:operatingSystemVersion:PlatformDescription', platformDescription_operatingSystemVersion - The version of the operating system used by the platform version.

$sel:programmingLanguages:PlatformDescription', platformDescription_programmingLanguages - The programming languages supported by the platform version.

$sel:solutionStackName:PlatformDescription', platformDescription_solutionStackName - The name of the solution stack used by the platform version.

$sel:platformName:PlatformDescription', platformDescription_platformName - The name of the platform version.

$sel:description:PlatformDescription', platformDescription_description - The description of the platform version.

$sel:supportedTierList:PlatformDescription', platformDescription_supportedTierList - The tiers supported by the platform version.

platformDescription_platformBranchName :: Lens' PlatformDescription (Maybe Text)

The platform branch to which the platform version belongs.

platformDescription_supportedAddonList :: Lens' PlatformDescription (Maybe [Text])

The additions supported by the platform version.

platformDescription_platformCategory :: Lens' PlatformDescription (Maybe Text)

The category of the platform version.

platformDescription_platformBranchLifecycleState :: Lens' PlatformDescription (Maybe Text)

The state of the platform version's branch in its lifecycle.

Possible values: Beta | Supported | Deprecated | Retired

platformDescription_platformVersion :: Lens' PlatformDescription (Maybe Text)

The version of the platform version.

platformDescription_platformStatus :: Lens' PlatformDescription (Maybe PlatformStatus)

The status of the platform version.

platformDescription_maintainer :: Lens' PlatformDescription (Maybe Text)

Information about the maintainer of the platform version.

platformDescription_platformLifecycleState :: Lens' PlatformDescription (Maybe Text)

The state of the platform version in its lifecycle.

Possible values: Recommended | null

If a null value is returned, the platform version isn't the recommended one for its branch. Each platform branch has a single recommended platform version, typically the most recent one.

platformDescription_platformOwner :: Lens' PlatformDescription (Maybe Text)

The AWS account ID of the person who created the platform version.

platformDescription_dateUpdated :: Lens' PlatformDescription (Maybe UTCTime)

The date when the platform version was last updated.

platformDescription_customAmiList :: Lens' PlatformDescription (Maybe [CustomAmi])

The custom AMIs supported by the platform version.

platformDescription_dateCreated :: Lens' PlatformDescription (Maybe UTCTime)

The date when the platform version was created.

platformDescription_operatingSystemName :: Lens' PlatformDescription (Maybe Text)

The operating system used by the platform version.

platformDescription_frameworks :: Lens' PlatformDescription (Maybe [PlatformFramework])

The frameworks supported by the platform version.

platformDescription_platformArn :: Lens' PlatformDescription (Maybe Text)

The ARN of the platform version.

platformDescription_operatingSystemVersion :: Lens' PlatformDescription (Maybe Text)

The version of the operating system used by the platform version.

platformDescription_programmingLanguages :: Lens' PlatformDescription (Maybe [PlatformProgrammingLanguage])

The programming languages supported by the platform version.

platformDescription_solutionStackName :: Lens' PlatformDescription (Maybe Text)

The name of the solution stack used by the platform version.

platformDescription_platformName :: Lens' PlatformDescription (Maybe Text)

The name of the platform version.

platformDescription_description :: Lens' PlatformDescription (Maybe Text)

The description of the platform version.

platformDescription_supportedTierList :: Lens' PlatformDescription (Maybe [Text])

The tiers supported by the platform version.