Amazonka.ElasticBeanstalk.UpdateTagsForResource

Description

Update the list of tags applied to an AWS Elastic Beanstalk resource. Two lists can be passed: TagsToAdd for tags to add or update, and TagsToRemove.

Elastic Beanstalk supports tagging of all of its resources. For details about resource tagging, see Tagging Application Resources.

If you create a custom IAM user policy to control permission to this operation, specify one of the following two virtual actions (or both) instead of the API operation name:

elasticbeanstalk:AddTags

Controls permission to call UpdateTagsForResource and pass a list of tags to add in the TagsToAdd parameter.

elasticbeanstalk:RemoveTags

Controls permission to call UpdateTagsForResource and pass a list of tag keys to remove in the TagsToRemove parameter.

For details about creating a custom user policy, see Creating a Custom User Policy.

Creating a Request

data UpdateTagsForResource

See: newUpdateTagsForResource smart constructor.

Constructors

UpdateTagsForResource'

Fields

tagsToRemove :: Maybe [Text]

A list of tag keys to remove. If a tag key doesn't exist, it is silently ignored.

Specify at least one of these parameters: TagsToAdd, TagsToRemove.

tagsToAdd :: Maybe [Tag]

A list of tags to add or update. If a key of an existing tag is added, the tag's value is updated.

Specify at least one of these parameters: TagsToAdd, TagsToRemove.

resourceArn :: Text

The Amazon Resource Name (ARN) of the resouce to be updated.

Must be the ARN of an Elastic Beanstalk resource.

newUpdateTagsForResource

Arguments

:: Text
$sel:resourceArn:UpdateTagsForResource'

-> UpdateTagsForResource

Create a value of UpdateTagsForResource with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:tagsToRemove:UpdateTagsForResource', updateTagsForResource_tagsToRemove - A list of tag keys to remove. If a tag key doesn't exist, it is silently ignored.

Specify at least one of these parameters: TagsToAdd, TagsToRemove.

$sel:tagsToAdd:UpdateTagsForResource', updateTagsForResource_tagsToAdd - A list of tags to add or update. If a key of an existing tag is added, the tag's value is updated.

Specify at least one of these parameters: TagsToAdd, TagsToRemove.

$sel:resourceArn:UpdateTagsForResource', updateTagsForResource_resourceArn - The Amazon Resource Name (ARN) of the resouce to be updated.

Must be the ARN of an Elastic Beanstalk resource.

Request Lenses

updateTagsForResource_tagsToRemove :: Lens' UpdateTagsForResource (Maybe [Text])

A list of tag keys to remove. If a tag key doesn't exist, it is silently ignored.

Specify at least one of these parameters: TagsToAdd, TagsToRemove.

updateTagsForResource_tagsToAdd :: Lens' UpdateTagsForResource (Maybe [Tag])

A list of tags to add or update. If a key of an existing tag is added, the tag's value is updated.

Specify at least one of these parameters: TagsToAdd, TagsToRemove.

updateTagsForResource_resourceArn :: Lens' UpdateTagsForResource Text

The Amazon Resource Name (ARN) of the resouce to be updated.

Must be the ARN of an Elastic Beanstalk resource.

Destructuring the Response

data UpdateTagsForResourceResponse

See: newUpdateTagsForResourceResponse smart constructor.

Constructors

UpdateTagsForResourceResponse'

newUpdateTagsForResourceResponse :: UpdateTagsForResourceResponse

Create a value of UpdateTagsForResourceResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.