Amazonka.ElasticBeanstalk.ValidateConfigurationSettings

Description

Takes a set of configuration settings and either a configuration template or environment, and determines whether those values are valid.

This action returns a list of messages indicating any errors or warnings associated with the selection of option values.

Creating a Request

data ValidateConfigurationSettings

A list of validation messages for a specified configuration template.

See: newValidateConfigurationSettings smart constructor.

Constructors

ValidateConfigurationSettings'

Fields

templateName :: Maybe Text

The name of the configuration template to validate the settings against.

Condition: You cannot specify both this and an environment name.

environmentName :: Maybe Text

The name of the environment to validate the settings against.

Condition: You cannot specify both this and a configuration template name.

applicationName :: Text

The name of the application that the configuration template or environment belongs to.

optionSettings :: [ConfigurationOptionSetting]

A list of the options and desired values to evaluate.

newValidateConfigurationSettings

Arguments

:: Text
$sel:applicationName:ValidateConfigurationSettings'

-> ValidateConfigurationSettings

Create a value of ValidateConfigurationSettings with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:templateName:ValidateConfigurationSettings', validateConfigurationSettings_templateName - The name of the configuration template to validate the settings against.

Condition: You cannot specify both this and an environment name.

$sel:environmentName:ValidateConfigurationSettings', validateConfigurationSettings_environmentName - The name of the environment to validate the settings against.

Condition: You cannot specify both this and a configuration template name.

$sel:applicationName:ValidateConfigurationSettings', validateConfigurationSettings_applicationName - The name of the application that the configuration template or environment belongs to.

$sel:optionSettings:ValidateConfigurationSettings', validateConfigurationSettings_optionSettings - A list of the options and desired values to evaluate.

Request Lenses

validateConfigurationSettings_templateName :: Lens' ValidateConfigurationSettings (Maybe Text)

The name of the configuration template to validate the settings against.

Condition: You cannot specify both this and an environment name.

validateConfigurationSettings_environmentName :: Lens' ValidateConfigurationSettings (Maybe Text)

The name of the environment to validate the settings against.

Condition: You cannot specify both this and a configuration template name.

validateConfigurationSettings_applicationName :: Lens' ValidateConfigurationSettings Text

The name of the application that the configuration template or environment belongs to.

validateConfigurationSettings_optionSettings :: Lens' ValidateConfigurationSettings [ConfigurationOptionSetting]

A list of the options and desired values to evaluate.

Destructuring the Response

data ValidateConfigurationSettingsResponse

Provides a list of validation messages.

See: newValidateConfigurationSettingsResponse smart constructor.

Constructors

ValidateConfigurationSettingsResponse'

Fields

messages :: Maybe [ValidationMessage]

A list of ValidationMessage.

httpStatus :: Int

The response's http status code.

newValidateConfigurationSettingsResponse

Arguments

:: Int
$sel:httpStatus:ValidateConfigurationSettingsResponse'

-> ValidateConfigurationSettingsResponse

Create a value of ValidateConfigurationSettingsResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:messages:ValidateConfigurationSettingsResponse', validateConfigurationSettingsResponse_messages - A list of ValidationMessage.

$sel:httpStatus:ValidateConfigurationSettingsResponse', validateConfigurationSettingsResponse_httpStatus - The response's http status code.

Response Lenses

validateConfigurationSettingsResponse_messages :: Lens' ValidateConfigurationSettingsResponse (Maybe [ValidationMessage])

A list of ValidationMessage.

validateConfigurationSettingsResponse_httpStatus :: Lens' ValidateConfigurationSettingsResponse Int

The response's http status code.