Amazonka.ElasticInference

Description

Derived from API version 2017-07-25 of the AWS service descriptions, licensed under Apache 2.0.

Elastic Inference public APIs.

Service Configuration

defaultService :: Service

API version 2017-07-25 of the Amazon Elastic Inference SDK configuration.

Errors

Error matchers are designed for use with the functions provided by Control.Exception.Lens. This allows catching (and rethrowing) service specific errors returned by ElasticInference.

InternalServerException

_InternalServerException :: AsError a => Getting (First ServiceError) a ServiceError

Raised when an unexpected error occurred during request processing.

ResourceNotFoundException

_ResourceNotFoundException :: AsError a => Getting (First ServiceError) a ServiceError

Raised when the requested resource cannot be found.

BadRequestException

_BadRequestException :: AsError a => Getting (First ServiceError) a ServiceError

Raised when a malformed input has been provided to the API.

Waiters

Waiters poll by repeatedly sending a request until some remote success condition configured by the Wait specification is fulfilled. The Wait specification determines how many attempts should be made, in addition to delay and retry strategies.

Operations

Some AWS operations return results that are incomplete and require subsequent requests in order to obtain the entire result set. The process of sending subsequent requests to continue where a previous request left off is called pagination. For example, the ListObjects operation of Amazon S3 returns up to 1000 objects at a time, and you must send subsequent requests with the appropriate Marker in order to retrieve the next page of results.

Operations that have an AWSPager instance can transparently perform subsequent requests, correctly setting Markers and other request facets to iterate through the entire result set of a truncated API operation. Operations which support this have an additional note in the documentation.

Many operations have the ability to filter results on the server side. See the individual operation parameters for details.

ListTagsForResource

data ListTagsForResource

See: newListTagsForResource smart constructor.

Constructors

ListTagsForResource' Text

newListTagsForResource

Arguments

:: Text
$sel:resourceArn:ListTagsForResource'

-> ListTagsForResource

Create a value of ListTagsForResource with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:resourceArn:ListTagsForResource', listTagsForResource_resourceArn - The ARN of the Elastic Inference Accelerator to list the tags for.

data ListTagsForResourceResponse

See: newListTagsForResourceResponse smart constructor.

Constructors

ListTagsForResourceResponse' (Maybe (HashMap Text Text)) Int

newListTagsForResourceResponse

Arguments

:: Int
$sel:httpStatus:ListTagsForResourceResponse'

-> ListTagsForResourceResponse

Create a value of ListTagsForResourceResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:tags:ListTagsForResourceResponse', listTagsForResourceResponse_tags - The tags of the Elastic Inference Accelerator.

$sel:httpStatus:ListTagsForResourceResponse', listTagsForResourceResponse_httpStatus - The response's http status code.

DescribeAcceleratorOfferings

data DescribeAcceleratorOfferings

See: newDescribeAcceleratorOfferings smart constructor.

Constructors

DescribeAcceleratorOfferings' (Maybe [Text]) LocationType

newDescribeAcceleratorOfferings

Arguments

:: LocationType
$sel:locationType:DescribeAcceleratorOfferings'

-> DescribeAcceleratorOfferings

Create a value of DescribeAcceleratorOfferings with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:acceleratorTypes:DescribeAcceleratorOfferings', describeAcceleratorOfferings_acceleratorTypes - The list of accelerator types to describe.

$sel:locationType:DescribeAcceleratorOfferings', describeAcceleratorOfferings_locationType - The location type that you want to describe accelerator type offerings for. It can assume the following values: region: will return the accelerator type offering at the regional level. availability-zone: will return the accelerator type offering at the availability zone level. availability-zone-id: will return the accelerator type offering at the availability zone level returning the availability zone id.

data DescribeAcceleratorOfferingsResponse

See: newDescribeAcceleratorOfferingsResponse smart constructor.

Constructors

DescribeAcceleratorOfferingsResponse' (Maybe [AcceleratorTypeOffering]) Int

newDescribeAcceleratorOfferingsResponse

Arguments

:: Int
$sel:httpStatus:DescribeAcceleratorOfferingsResponse'

-> DescribeAcceleratorOfferingsResponse

Create a value of DescribeAcceleratorOfferingsResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:acceleratorTypeOfferings:DescribeAcceleratorOfferingsResponse', describeAcceleratorOfferingsResponse_acceleratorTypeOfferings - The list of accelerator type offerings for a specific location.

$sel:httpStatus:DescribeAcceleratorOfferingsResponse', describeAcceleratorOfferingsResponse_httpStatus - The response's http status code.

DescribeAccelerators (Paginated)

data DescribeAccelerators

See: newDescribeAccelerators smart constructor.

Constructors

DescribeAccelerators' (Maybe [Filter]) (Maybe Text) (Maybe Natural) (Maybe [Text])

newDescribeAccelerators :: DescribeAccelerators

Create a value of DescribeAccelerators with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:filters:DescribeAccelerators', describeAccelerators_filters - One or more filters. Filter names and values are case-sensitive. Valid filter names are: accelerator-types: can provide a list of accelerator type names to filter for. instance-id: can provide a list of EC2 instance ids to filter for.

$sel:nextToken:DescribeAccelerators', describeAccelerators_nextToken - A token to specify where to start paginating. This is the NextToken from a previously truncated response.

$sel:maxResults:DescribeAccelerators', describeAccelerators_maxResults - The total number of items to return in the command's output. If the total number of items available is more than the value specified, a NextToken is provided in the command's output. To resume pagination, provide the NextToken value in the starting-token argument of a subsequent command. Do not use the NextToken response element directly outside of the AWS CLI.

$sel:acceleratorIds:DescribeAccelerators', describeAccelerators_acceleratorIds - The IDs of the accelerators to describe.

data DescribeAcceleratorsResponse

See: newDescribeAcceleratorsResponse smart constructor.

Constructors

DescribeAcceleratorsResponse' (Maybe [ElasticInferenceAccelerator]) (Maybe Text) Int

newDescribeAcceleratorsResponse

Arguments

:: Int
$sel:httpStatus:DescribeAcceleratorsResponse'

-> DescribeAcceleratorsResponse

Create a value of DescribeAcceleratorsResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:acceleratorSet:DescribeAcceleratorsResponse', describeAcceleratorsResponse_acceleratorSet - The details of the Elastic Inference Accelerators.

$sel:nextToken:DescribeAccelerators', describeAcceleratorsResponse_nextToken - A token to specify where to start paginating. This is the NextToken from a previously truncated response.

$sel:httpStatus:DescribeAcceleratorsResponse', describeAcceleratorsResponse_httpStatus - The response's http status code.

TagResource

data TagResource

See: newTagResource smart constructor.

Constructors

TagResource' Text (HashMap Text Text)

newTagResource

Arguments

:: Text
$sel:resourceArn:TagResource'

-> TagResource

Create a value of TagResource with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:resourceArn:TagResource', tagResource_resourceArn - The ARN of the Elastic Inference Accelerator to tag.

$sel:tags:TagResource', tagResource_tags - The tags to add to the Elastic Inference Accelerator.

data TagResourceResponse

See: newTagResourceResponse smart constructor.

Constructors

TagResourceResponse' Int

newTagResourceResponse

Arguments

:: Int
$sel:httpStatus:TagResourceResponse'

-> TagResourceResponse

Create a value of TagResourceResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:httpStatus:TagResourceResponse', tagResourceResponse_httpStatus - The response's http status code.

UntagResource

data UntagResource

See: newUntagResource smart constructor.

Constructors

UntagResource' Text (NonEmpty Text)

newUntagResource

Arguments

:: Text
$sel:resourceArn:UntagResource'

-> NonEmpty Text
$sel:tagKeys:UntagResource'

-> UntagResource

Create a value of UntagResource with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:resourceArn:UntagResource', untagResource_resourceArn - The ARN of the Elastic Inference Accelerator to untag.

$sel:tagKeys:UntagResource', untagResource_tagKeys - The list of tags to remove from the Elastic Inference Accelerator.

data UntagResourceResponse

See: newUntagResourceResponse smart constructor.

Constructors

UntagResourceResponse' Int

newUntagResourceResponse

Arguments

:: Int
$sel:httpStatus:UntagResourceResponse'

-> UntagResourceResponse

Create a value of UntagResourceResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:httpStatus:UntagResourceResponse', untagResourceResponse_httpStatus - The response's http status code.

DescribeAcceleratorTypes

data DescribeAcceleratorTypes

See: newDescribeAcceleratorTypes smart constructor.

Constructors

DescribeAcceleratorTypes'

newDescribeAcceleratorTypes :: DescribeAcceleratorTypes

Create a value of DescribeAcceleratorTypes with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

data DescribeAcceleratorTypesResponse

See: newDescribeAcceleratorTypesResponse smart constructor.

Constructors

DescribeAcceleratorTypesResponse' (Maybe [AcceleratorType]) Int

newDescribeAcceleratorTypesResponse

Arguments

:: Int
$sel:httpStatus:DescribeAcceleratorTypesResponse'

-> DescribeAcceleratorTypesResponse

Create a value of DescribeAcceleratorTypesResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:acceleratorTypes:DescribeAcceleratorTypesResponse', describeAcceleratorTypesResponse_acceleratorTypes - The available accelerator types.

$sel:httpStatus:DescribeAcceleratorTypesResponse', describeAcceleratorTypesResponse_httpStatus - The response's http status code.

Types

LocationType

newtype LocationType

Constructors

LocationType'

Fields

fromLocationType :: Text

Bundled Patterns

pattern LocationType_Availability_zone :: LocationType

pattern LocationType_Availability_zone_id :: LocationType

pattern LocationType_Region :: LocationType

AcceleratorType

data AcceleratorType

The details of an Elastic Inference Accelerator type.

See: newAcceleratorType smart constructor.

Constructors

AcceleratorType' (Maybe [KeyValuePair]) (Maybe MemoryInfo) (Maybe Text)

newAcceleratorType :: AcceleratorType

Create a value of AcceleratorType with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:throughputInfo:AcceleratorType', acceleratorType_throughputInfo - The throughput information of the Elastic Inference Accelerator type.

$sel:memoryInfo:AcceleratorType', acceleratorType_memoryInfo - The memory information of the Elastic Inference Accelerator type.

$sel:acceleratorTypeName:AcceleratorType', acceleratorType_acceleratorTypeName - The name of the Elastic Inference Accelerator type.

AcceleratorTypeOffering

data AcceleratorTypeOffering

The offering for an Elastic Inference Accelerator type.

See: newAcceleratorTypeOffering smart constructor.

Constructors

AcceleratorTypeOffering' (Maybe Text) (Maybe Text) (Maybe LocationType)

newAcceleratorTypeOffering :: AcceleratorTypeOffering

Create a value of AcceleratorTypeOffering with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:acceleratorType:AcceleratorTypeOffering', acceleratorTypeOffering_acceleratorType - The name of the Elastic Inference Accelerator type.

$sel:location:AcceleratorTypeOffering', acceleratorTypeOffering_location - The location for the offering. It will return either the region, availability zone or availability zone id for the offering depending on the locationType value.

$sel:locationType:AcceleratorTypeOffering', acceleratorTypeOffering_locationType - The location type for the offering. It can assume the following values: region: defines that the offering is at the regional level. availability-zone: defines that the offering is at the availability zone level. availability-zone-id: defines that the offering is at the availability zone level, defined by the availability zone id.

ElasticInferenceAccelerator

data ElasticInferenceAccelerator

The details of an Elastic Inference Accelerator.

See: newElasticInferenceAccelerator smart constructor.

Constructors

ElasticInferenceAccelerator' (Maybe Text) (Maybe Text) (Maybe Text) (Maybe ElasticInferenceAcceleratorHealth) (Maybe Text)

newElasticInferenceAccelerator :: ElasticInferenceAccelerator

Create a value of ElasticInferenceAccelerator with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:acceleratorType:ElasticInferenceAccelerator', elasticInferenceAccelerator_acceleratorType - The type of the Elastic Inference Accelerator.

$sel:acceleratorId:ElasticInferenceAccelerator', elasticInferenceAccelerator_acceleratorId - The ID of the Elastic Inference Accelerator.

$sel:attachedResource:ElasticInferenceAccelerator', elasticInferenceAccelerator_attachedResource - The ARN of the resource that the Elastic Inference Accelerator is attached to.

$sel:acceleratorHealth:ElasticInferenceAccelerator', elasticInferenceAccelerator_acceleratorHealth - The health of the Elastic Inference Accelerator.

$sel:availabilityZone:ElasticInferenceAccelerator', elasticInferenceAccelerator_availabilityZone - The availability zone where the Elastic Inference Accelerator is present.

ElasticInferenceAcceleratorHealth

data ElasticInferenceAcceleratorHealth

The health details of an Elastic Inference Accelerator.

See: newElasticInferenceAcceleratorHealth smart constructor.

Constructors

ElasticInferenceAcceleratorHealth' (Maybe Text)

newElasticInferenceAcceleratorHealth :: ElasticInferenceAcceleratorHealth

Create a value of ElasticInferenceAcceleratorHealth with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:status:ElasticInferenceAcceleratorHealth', elasticInferenceAcceleratorHealth_status - The health status of the Elastic Inference Accelerator.

Filter

data Filter

A filter expression for the Elastic Inference Accelerator list.

See: newFilter smart constructor.

Constructors

Filter' (Maybe [Text]) (Maybe Text)

newFilter :: Filter

Create a value of Filter with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:values:Filter', filter_values - The values for the filter of the Elastic Inference Accelerator list.

$sel:name:Filter', filter_name - The filter name for the Elastic Inference Accelerator list. It can assume the following values: accelerator-type: the type of Elastic Inference Accelerator to filter for. instance-id: an EC2 instance id to filter for.

KeyValuePair

data KeyValuePair

A throughput entry for an Elastic Inference Accelerator type.

See: newKeyValuePair smart constructor.

Constructors

KeyValuePair' (Maybe Int) (Maybe Text)

newKeyValuePair :: KeyValuePair

Create a value of KeyValuePair with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:value:KeyValuePair', keyValuePair_value - The throughput value of the Elastic Inference Accelerator type.

$sel:key:KeyValuePair', keyValuePair_key - The throughput value of the Elastic Inference Accelerator type. It can assume the following values: TFLOPS16bit: the throughput expressed in 16bit TeraFLOPS. TFLOPS32bit: the throughput expressed in 32bit TeraFLOPS.

MemoryInfo

data MemoryInfo

The memory information of an Elastic Inference Accelerator type.

See: newMemoryInfo smart constructor.

Constructors

MemoryInfo' (Maybe Int)

newMemoryInfo :: MemoryInfo

Create a value of MemoryInfo with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:sizeInMiB:MemoryInfo', memoryInfo_sizeInMiB - The size in mebibytes of the Elastic Inference Accelerator type.