Amazonka.ElasticInference.DescribeAccelerators

Description

Describes information over a provided set of accelerators belonging to an account.

This operation returns paginated results.

Creating a Request

data DescribeAccelerators

See: newDescribeAccelerators smart constructor.

Constructors

DescribeAccelerators'

Fields

filters :: Maybe [Filter]

One or more filters. Filter names and values are case-sensitive. Valid filter names are: accelerator-types: can provide a list of accelerator type names to filter for. instance-id: can provide a list of EC2 instance ids to filter for.

nextToken :: Maybe Text

A token to specify where to start paginating. This is the NextToken from a previously truncated response.

maxResults :: Maybe Natural

The total number of items to return in the command's output. If the total number of items available is more than the value specified, a NextToken is provided in the command's output. To resume pagination, provide the NextToken value in the starting-token argument of a subsequent command. Do not use the NextToken response element directly outside of the AWS CLI.

acceleratorIds :: Maybe [Text]

The IDs of the accelerators to describe.

newDescribeAccelerators :: DescribeAccelerators

Create a value of DescribeAccelerators with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:filters:DescribeAccelerators', describeAccelerators_filters - One or more filters. Filter names and values are case-sensitive. Valid filter names are: accelerator-types: can provide a list of accelerator type names to filter for. instance-id: can provide a list of EC2 instance ids to filter for.

$sel:nextToken:DescribeAccelerators', describeAccelerators_nextToken - A token to specify where to start paginating. This is the NextToken from a previously truncated response.

$sel:maxResults:DescribeAccelerators', describeAccelerators_maxResults - The total number of items to return in the command's output. If the total number of items available is more than the value specified, a NextToken is provided in the command's output. To resume pagination, provide the NextToken value in the starting-token argument of a subsequent command. Do not use the NextToken response element directly outside of the AWS CLI.

$sel:acceleratorIds:DescribeAccelerators', describeAccelerators_acceleratorIds - The IDs of the accelerators to describe.

Request Lenses

describeAccelerators_filters :: Lens' DescribeAccelerators (Maybe [Filter])

One or more filters. Filter names and values are case-sensitive. Valid filter names are: accelerator-types: can provide a list of accelerator type names to filter for. instance-id: can provide a list of EC2 instance ids to filter for.

describeAccelerators_nextToken :: Lens' DescribeAccelerators (Maybe Text)

A token to specify where to start paginating. This is the NextToken from a previously truncated response.

describeAccelerators_maxResults :: Lens' DescribeAccelerators (Maybe Natural)

The total number of items to return in the command's output. If the total number of items available is more than the value specified, a NextToken is provided in the command's output. To resume pagination, provide the NextToken value in the starting-token argument of a subsequent command. Do not use the NextToken response element directly outside of the AWS CLI.

describeAccelerators_acceleratorIds :: Lens' DescribeAccelerators (Maybe [Text])

The IDs of the accelerators to describe.

Destructuring the Response

data DescribeAcceleratorsResponse

See: newDescribeAcceleratorsResponse smart constructor.

Constructors

DescribeAcceleratorsResponse'

Fields

acceleratorSet :: Maybe [ElasticInferenceAccelerator]

The details of the Elastic Inference Accelerators.

nextToken :: Maybe Text

A token to specify where to start paginating. This is the NextToken from a previously truncated response.

httpStatus :: Int

The response's http status code.

newDescribeAcceleratorsResponse

Arguments

:: Int
$sel:httpStatus:DescribeAcceleratorsResponse'

-> DescribeAcceleratorsResponse

Create a value of DescribeAcceleratorsResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:acceleratorSet:DescribeAcceleratorsResponse', describeAcceleratorsResponse_acceleratorSet - The details of the Elastic Inference Accelerators.

$sel:nextToken:DescribeAccelerators', describeAcceleratorsResponse_nextToken - A token to specify where to start paginating. This is the NextToken from a previously truncated response.

$sel:httpStatus:DescribeAcceleratorsResponse', describeAcceleratorsResponse_httpStatus - The response's http status code.

Response Lenses

describeAcceleratorsResponse_acceleratorSet :: Lens' DescribeAcceleratorsResponse (Maybe [ElasticInferenceAccelerator])

The details of the Elastic Inference Accelerators.

describeAcceleratorsResponse_nextToken :: Lens' DescribeAcceleratorsResponse (Maybe Text)

A token to specify where to start paginating. This is the NextToken from a previously truncated response.

describeAcceleratorsResponse_httpStatus :: Lens' DescribeAcceleratorsResponse Int

The response's http status code.