Amazonka.ElasticInference.Types.AcceleratorType

Description

Documentation

data AcceleratorType

The details of an Elastic Inference Accelerator type.

See: newAcceleratorType smart constructor.

Constructors

AcceleratorType'

Fields

throughputInfo :: Maybe [KeyValuePair]

The throughput information of the Elastic Inference Accelerator type.

memoryInfo :: Maybe MemoryInfo

The memory information of the Elastic Inference Accelerator type.

acceleratorTypeName :: Maybe Text

The name of the Elastic Inference Accelerator type.

newAcceleratorType :: AcceleratorType

Create a value of AcceleratorType with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:throughputInfo:AcceleratorType', acceleratorType_throughputInfo - The throughput information of the Elastic Inference Accelerator type.

$sel:memoryInfo:AcceleratorType', acceleratorType_memoryInfo - The memory information of the Elastic Inference Accelerator type.

$sel:acceleratorTypeName:AcceleratorType', acceleratorType_acceleratorTypeName - The name of the Elastic Inference Accelerator type.

acceleratorType_throughputInfo :: Lens' AcceleratorType (Maybe [KeyValuePair])

The throughput information of the Elastic Inference Accelerator type.

acceleratorType_memoryInfo :: Lens' AcceleratorType (Maybe MemoryInfo)

The memory information of the Elastic Inference Accelerator type.

acceleratorType_acceleratorTypeName :: Lens' AcceleratorType (Maybe Text)

The name of the Elastic Inference Accelerator type.