Amazonka.ElasticSearch.CreatePackage

Description

Create a package for use with Amazon ES domains.

Creating a Request

data CreatePackage

Container for request parameters to CreatePackage operation.

See: newCreatePackage smart constructor.

Constructors

CreatePackage'

Fields

packageDescription :: Maybe Text

Description of the package.

packageName :: Text

Unique identifier for the package.

packageType :: PackageType

Type of package. Currently supports only TXT-DICTIONARY.

packageSource :: PackageSource

The customer S3 location PackageSource for importing the package.

newCreatePackage

Arguments

:: Text
$sel:packageName:CreatePackage'

-> PackageType
$sel:packageType:CreatePackage'

-> PackageSource
$sel:packageSource:CreatePackage'

-> CreatePackage

Create a value of CreatePackage with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:packageDescription:CreatePackage', createPackage_packageDescription - Description of the package.

$sel:packageName:CreatePackage', createPackage_packageName - Unique identifier for the package.

$sel:packageType:CreatePackage', createPackage_packageType - Type of package. Currently supports only TXT-DICTIONARY.

$sel:packageSource:CreatePackage', createPackage_packageSource - The customer S3 location PackageSource for importing the package.

Request Lenses

createPackage_packageDescription :: Lens' CreatePackage (Maybe Text)

Description of the package.

createPackage_packageName :: Lens' CreatePackage Text

Unique identifier for the package.

createPackage_packageType :: Lens' CreatePackage PackageType

Type of package. Currently supports only TXT-DICTIONARY.

createPackage_packageSource :: Lens' CreatePackage PackageSource

The customer S3 location PackageSource for importing the package.

Destructuring the Response

data CreatePackageResponse

Container for response returned by CreatePackage operation.

See: newCreatePackageResponse smart constructor.

Constructors

CreatePackageResponse'

Fields

packageDetails :: Maybe PackageDetails

Information about the package PackageDetails.

httpStatus :: Int

The response's http status code.

newCreatePackageResponse

Arguments

:: Int
$sel:httpStatus:CreatePackageResponse'

-> CreatePackageResponse

Create a value of CreatePackageResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:packageDetails:CreatePackageResponse', createPackageResponse_packageDetails - Information about the package PackageDetails.

$sel:httpStatus:CreatePackageResponse', createPackageResponse_httpStatus - The response's http status code.

Response Lenses

createPackageResponse_packageDetails :: Lens' CreatePackageResponse (Maybe PackageDetails)

Information about the package PackageDetails.

createPackageResponse_httpStatus :: Lens' CreatePackageResponse Int

The response's http status code.