Amazonka.ElasticSearch.ListElasticsearchVersions

Description

List all supported Elasticsearch versions

This operation returns paginated results.

Creating a Request

data ListElasticsearchVersions

Container for the parameters to the ListElasticsearchVersions operation.

Use MaxResults to control the maximum number of results to retrieve in a single call.

Use NextToken in response to retrieve more results. If the received response does not contain a NextToken, then there are no more results to retrieve.

See: newListElasticsearchVersions smart constructor.

Constructors

ListElasticsearchVersions'

Fields

nextToken :: Maybe Text

maxResults :: Maybe Int

Set this value to limit the number of results returned. Value provided must be greater than 10 else it wont be honored.

newListElasticsearchVersions :: ListElasticsearchVersions

Create a value of ListElasticsearchVersions with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:nextToken:ListElasticsearchVersions', listElasticsearchVersions_nextToken - Undocumented member.

$sel:maxResults:ListElasticsearchVersions', listElasticsearchVersions_maxResults - Set this value to limit the number of results returned. Value provided must be greater than 10 else it wont be honored.

Request Lenses

listElasticsearchVersions_nextToken :: Lens' ListElasticsearchVersions (Maybe Text)

Undocumented member.

listElasticsearchVersions_maxResults :: Lens' ListElasticsearchVersions (Maybe Int)

Set this value to limit the number of results returned. Value provided must be greater than 10 else it wont be honored.

Destructuring the Response

data ListElasticsearchVersionsResponse

Container for the parameters for response received from ListElasticsearchVersions operation.

See: newListElasticsearchVersionsResponse smart constructor.

Constructors

ListElasticsearchVersionsResponse'

Fields

nextToken :: Maybe Text

elasticsearchVersions :: Maybe [Text]

httpStatus :: Int

The response's http status code.

newListElasticsearchVersionsResponse

Arguments

:: Int
$sel:httpStatus:ListElasticsearchVersionsResponse'

-> ListElasticsearchVersionsResponse

Create a value of ListElasticsearchVersionsResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:nextToken:ListElasticsearchVersions', listElasticsearchVersionsResponse_nextToken - Undocumented member.

$sel:elasticsearchVersions:ListElasticsearchVersionsResponse', listElasticsearchVersionsResponse_elasticsearchVersions - Undocumented member.

$sel:httpStatus:ListElasticsearchVersionsResponse', listElasticsearchVersionsResponse_httpStatus - The response's http status code.

Response Lenses

listElasticsearchVersionsResponse_nextToken :: Lens' ListElasticsearchVersionsResponse (Maybe Text)

Undocumented member.

listElasticsearchVersionsResponse_elasticsearchVersions :: Lens' ListElasticsearchVersionsResponse (Maybe [Text])

Undocumented member.

listElasticsearchVersionsResponse_httpStatus :: Lens' ListElasticsearchVersionsResponse Int

The response's http status code.