Amazonka.ElasticSearch.PurchaseReservedElasticsearchInstanceOffering

Description

Allows you to purchase reserved Elasticsearch instances.

Creating a Request

data PurchaseReservedElasticsearchInstanceOffering

Container for parameters to PurchaseReservedElasticsearchInstanceOffering

See: newPurchaseReservedElasticsearchInstanceOffering smart constructor.

Constructors

PurchaseReservedElasticsearchInstanceOffering'

Fields

instanceCount :: Maybe Natural

The number of Elasticsearch instances to reserve.

reservedElasticsearchInstanceOfferingId :: Text

The ID of the reserved Elasticsearch instance offering to purchase.

reservationName :: Text

A customer-specified identifier to track this reservation.

newPurchaseReservedElasticsearchInstanceOffering

Arguments

:: Text
$sel:reservedElasticsearchInstanceOfferingId:PurchaseReservedElasticsearchInstanceOffering'

-> Text
$sel:reservationName:PurchaseReservedElasticsearchInstanceOffering'

-> PurchaseReservedElasticsearchInstanceOffering

Create a value of PurchaseReservedElasticsearchInstanceOffering with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:instanceCount:PurchaseReservedElasticsearchInstanceOffering', purchaseReservedElasticsearchInstanceOffering_instanceCount - The number of Elasticsearch instances to reserve.

$sel:reservedElasticsearchInstanceOfferingId:PurchaseReservedElasticsearchInstanceOffering', purchaseReservedElasticsearchInstanceOffering_reservedElasticsearchInstanceOfferingId - The ID of the reserved Elasticsearch instance offering to purchase.

$sel:reservationName:PurchaseReservedElasticsearchInstanceOffering', purchaseReservedElasticsearchInstanceOffering_reservationName - A customer-specified identifier to track this reservation.

Request Lenses

purchaseReservedElasticsearchInstanceOffering_instanceCount :: Lens' PurchaseReservedElasticsearchInstanceOffering (Maybe Natural)

The number of Elasticsearch instances to reserve.

purchaseReservedElasticsearchInstanceOffering_reservedElasticsearchInstanceOfferingId :: Lens' PurchaseReservedElasticsearchInstanceOffering Text

The ID of the reserved Elasticsearch instance offering to purchase.

purchaseReservedElasticsearchInstanceOffering_reservationName :: Lens' PurchaseReservedElasticsearchInstanceOffering Text

A customer-specified identifier to track this reservation.

Destructuring the Response

data PurchaseReservedElasticsearchInstanceOfferingResponse

Represents the output of a PurchaseReservedElasticsearchInstanceOffering operation.

See: newPurchaseReservedElasticsearchInstanceOfferingResponse smart constructor.

Constructors

PurchaseReservedElasticsearchInstanceOfferingResponse'

Fields

reservedElasticsearchInstanceId :: Maybe Text

Details of the reserved Elasticsearch instance which was purchased.

reservationName :: Maybe Text

The customer-specified identifier used to track this reservation.

httpStatus :: Int

The response's http status code.

newPurchaseReservedElasticsearchInstanceOfferingResponse

Arguments

:: Int
$sel:httpStatus:PurchaseReservedElasticsearchInstanceOfferingResponse'

-> PurchaseReservedElasticsearchInstanceOfferingResponse

Create a value of PurchaseReservedElasticsearchInstanceOfferingResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:reservedElasticsearchInstanceId:PurchaseReservedElasticsearchInstanceOfferingResponse', purchaseReservedElasticsearchInstanceOfferingResponse_reservedElasticsearchInstanceId - Details of the reserved Elasticsearch instance which was purchased.

$sel:reservationName:PurchaseReservedElasticsearchInstanceOffering', purchaseReservedElasticsearchInstanceOfferingResponse_reservationName - The customer-specified identifier used to track this reservation.

$sel:httpStatus:PurchaseReservedElasticsearchInstanceOfferingResponse', purchaseReservedElasticsearchInstanceOfferingResponse_httpStatus - The response's http status code.

Response Lenses

purchaseReservedElasticsearchInstanceOfferingResponse_reservedElasticsearchInstanceId :: Lens' PurchaseReservedElasticsearchInstanceOfferingResponse (Maybe Text)

Details of the reserved Elasticsearch instance which was purchased.

purchaseReservedElasticsearchInstanceOfferingResponse_reservationName :: Lens' PurchaseReservedElasticsearchInstanceOfferingResponse (Maybe Text)

The customer-specified identifier used to track this reservation.

purchaseReservedElasticsearchInstanceOfferingResponse_httpStatus :: Lens' PurchaseReservedElasticsearchInstanceOfferingResponse Int

The response's http status code.