Amazonka.ElasticSearch.Types.AdvancedSecurityOptionsInput

Description

Documentation

data AdvancedSecurityOptionsInput

Specifies the advanced security configuration: whether advanced security is enabled, whether the internal database option is enabled, master username and password (if internal database is enabled), and master user ARN (if IAM is enabled).

See: newAdvancedSecurityOptionsInput smart constructor.

Constructors

AdvancedSecurityOptionsInput'

Fields

enabled :: Maybe Bool

True if advanced security is enabled.

internalUserDatabaseEnabled :: Maybe Bool

True if the internal user database is enabled.

masterUserOptions :: Maybe MasterUserOptions

Credentials for the master user: username and password, ARN, or both.

sAMLOptions :: Maybe SAMLOptionsInput

Specifies the SAML application configuration for the domain.

newAdvancedSecurityOptionsInput :: AdvancedSecurityOptionsInput

Create a value of AdvancedSecurityOptionsInput with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:enabled:AdvancedSecurityOptionsInput', advancedSecurityOptionsInput_enabled - True if advanced security is enabled.

$sel:internalUserDatabaseEnabled:AdvancedSecurityOptionsInput', advancedSecurityOptionsInput_internalUserDatabaseEnabled - True if the internal user database is enabled.

$sel:masterUserOptions:AdvancedSecurityOptionsInput', advancedSecurityOptionsInput_masterUserOptions - Credentials for the master user: username and password, ARN, or both.

$sel:sAMLOptions:AdvancedSecurityOptionsInput', advancedSecurityOptionsInput_sAMLOptions - Specifies the SAML application configuration for the domain.

advancedSecurityOptionsInput_enabled :: Lens' AdvancedSecurityOptionsInput (Maybe Bool)

True if advanced security is enabled.

advancedSecurityOptionsInput_internalUserDatabaseEnabled :: Lens' AdvancedSecurityOptionsInput (Maybe Bool)

True if the internal user database is enabled.

advancedSecurityOptionsInput_masterUserOptions :: Lens' AdvancedSecurityOptionsInput (Maybe MasterUserOptions)

Credentials for the master user: username and password, ARN, or both.

advancedSecurityOptionsInput_sAMLOptions :: Lens' AdvancedSecurityOptionsInput (Maybe SAMLOptionsInput)

Specifies the SAML application configuration for the domain.