Amazonka.ElasticSearch.Types.CompatibleVersionsMap

Description

Documentation

data CompatibleVersionsMap

A map from an ElasticsearchVersion to a list of compatible ElasticsearchVersion s to which the domain can be upgraded.

See: newCompatibleVersionsMap smart constructor.

Constructors

CompatibleVersionsMap'

Fields

sourceVersion :: Maybe Text

The current version of Elasticsearch on which a domain is.

targetVersions :: Maybe [Text]

newCompatibleVersionsMap :: CompatibleVersionsMap

Create a value of CompatibleVersionsMap with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:sourceVersion:CompatibleVersionsMap', compatibleVersionsMap_sourceVersion - The current version of Elasticsearch on which a domain is.

$sel:targetVersions:CompatibleVersionsMap', compatibleVersionsMap_targetVersions - Undocumented member.

compatibleVersionsMap_sourceVersion :: Lens' CompatibleVersionsMap (Maybe Text)

The current version of Elasticsearch on which a domain is.

compatibleVersionsMap_targetVersions :: Lens' CompatibleVersionsMap (Maybe [Text])

Undocumented member.