Amazonka.ElasticSearch.Types.StorageType

Description

Documentation

data StorageType

StorageTypes represents the list of storage related types and their attributes that are available for given InstanceType.

See: newStorageType smart constructor.

Constructors

StorageType'

Fields

storageTypeLimits :: Maybe [StorageTypeLimit]

List of limits that are applicable for given storage type.

storageSubTypeName :: Maybe Text

storageTypeName :: Maybe Text

newStorageType :: StorageType

Create a value of StorageType with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:storageTypeLimits:StorageType', storageType_storageTypeLimits - List of limits that are applicable for given storage type.

$sel:storageSubTypeName:StorageType', storageType_storageSubTypeName - Undocumented member.

$sel:storageTypeName:StorageType', storageType_storageTypeName - Undocumented member.

storageType_storageTypeLimits :: Lens' StorageType (Maybe [StorageTypeLimit])

List of limits that are applicable for given storage type.

storageType_storageSubTypeName :: Lens' StorageType (Maybe Text)

Undocumented member.

storageType_storageTypeName :: Lens' StorageType (Maybe Text)

Undocumented member.