Amazonka.ElasticSearch.Types.UpgradeHistory

Description

Documentation

data UpgradeHistory

History of the last 10 Upgrades and Upgrade Eligibility Checks.

See: newUpgradeHistory smart constructor.

Constructors

UpgradeHistory'

Fields

upgradeStatus :: Maybe UpgradeStatus

The overall status of the update. The status can take one of the following values:

In Progress

Succeeded

Succeeded with Issues

Failed

stepsList :: Maybe [UpgradeStepItem]

A list of UpgradeStepItem s representing information about each step performed as pard of a specific Upgrade or Upgrade Eligibility Check.

upgradeName :: Maybe Text

A string that describes the update briefly

startTimestamp :: Maybe POSIX

UTC Timestamp at which the Upgrade API call was made in "yyyy-MM-ddTHH:mm:ssZ" format.

newUpgradeHistory :: UpgradeHistory

Create a value of UpgradeHistory with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:upgradeStatus:UpgradeHistory', upgradeHistory_upgradeStatus - The overall status of the update. The status can take one of the following values:

In Progress

Succeeded

Succeeded with Issues

Failed

$sel:stepsList:UpgradeHistory', upgradeHistory_stepsList - A list of UpgradeStepItem s representing information about each step performed as pard of a specific Upgrade or Upgrade Eligibility Check.

$sel:upgradeName:UpgradeHistory', upgradeHistory_upgradeName - A string that describes the update briefly

$sel:startTimestamp:UpgradeHistory', upgradeHistory_startTimestamp - UTC Timestamp at which the Upgrade API call was made in "yyyy-MM-ddTHH:mm:ssZ" format.

upgradeHistory_upgradeStatus :: Lens' UpgradeHistory (Maybe UpgradeStatus)

The overall status of the update. The status can take one of the following values:

In Progress

Succeeded

Succeeded with Issues

Failed

upgradeHistory_stepsList :: Lens' UpgradeHistory (Maybe [UpgradeStepItem])

A list of UpgradeStepItem s representing information about each step performed as pard of a specific Upgrade or Upgrade Eligibility Check.

upgradeHistory_upgradeName :: Lens' UpgradeHistory (Maybe Text)

A string that describes the update briefly

upgradeHistory_startTimestamp :: Lens' UpgradeHistory (Maybe UTCTime)

UTC Timestamp at which the Upgrade API call was made in "yyyy-MM-ddTHH:mm:ssZ" format.