Amazonka.ElasticSearch.UpdateElasticsearchDomainConfig

Description

Modifies the cluster configuration of the specified Elasticsearch domain, setting as setting the instance type and the number of instances.

Creating a Request

data UpdateElasticsearchDomainConfig

Container for the parameters to the UpdateElasticsearchDomain operation. Specifies the type and number of instances in the domain cluster.

See: newUpdateElasticsearchDomainConfig smart constructor.

Constructors

UpdateElasticsearchDomainConfig'

Fields

eBSOptions :: Maybe EBSOptions

Specify the type and size of the EBS volume that you want to use.

nodeToNodeEncryptionOptions :: Maybe NodeToNodeEncryptionOptions

Specifies the NodeToNodeEncryptionOptions.

accessPolicies :: Maybe Text

IAM access policy as a JSON-formatted string.

autoTuneOptions :: Maybe AutoTuneOptions

Specifies Auto-Tune options.

logPublishingOptions :: Maybe (HashMap LogType LogPublishingOption)

Map of LogType and LogPublishingOption, each containing options to publish a given type of Elasticsearch log.

advancedSecurityOptions :: Maybe AdvancedSecurityOptionsInput

Specifies advanced security options.

elasticsearchClusterConfig :: Maybe ElasticsearchClusterConfig

The type and number of instances to instantiate for the domain cluster.

snapshotOptions :: Maybe SnapshotOptions

Option to set the time, in UTC format, for the daily automated snapshot. Default value is 0 hours.

cognitoOptions :: Maybe CognitoOptions

Options to specify the Cognito user and identity pools for Kibana authentication. For more information, see Amazon Cognito Authentication for Kibana.

encryptionAtRestOptions :: Maybe EncryptionAtRestOptions

Specifies the Encryption At Rest Options.

vPCOptions :: Maybe VPCOptions

Options to specify the subnets and security groups for VPC endpoint. For more information, see Creating a VPC in VPC Endpoints for Amazon Elasticsearch Service Domains

domainEndpointOptions :: Maybe DomainEndpointOptions

Options to specify configuration that will be applied to the domain endpoint.

advancedOptions :: Maybe (HashMap Text Text)

Modifies the advanced option to allow references to indices in an HTTP request body. Must be false when configuring access to individual sub-resources. By default, the value is true. See Configuration Advanced Options for more information.

domainName :: Text

The name of the Elasticsearch domain that you are updating.

newUpdateElasticsearchDomainConfig

Arguments

:: Text
$sel:domainName:UpdateElasticsearchDomainConfig'

-> UpdateElasticsearchDomainConfig

Create a value of UpdateElasticsearchDomainConfig with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:eBSOptions:UpdateElasticsearchDomainConfig', updateElasticsearchDomainConfig_eBSOptions - Specify the type and size of the EBS volume that you want to use.

$sel:nodeToNodeEncryptionOptions:UpdateElasticsearchDomainConfig', updateElasticsearchDomainConfig_nodeToNodeEncryptionOptions - Specifies the NodeToNodeEncryptionOptions.

$sel:accessPolicies:UpdateElasticsearchDomainConfig', updateElasticsearchDomainConfig_accessPolicies - IAM access policy as a JSON-formatted string.

$sel:autoTuneOptions:UpdateElasticsearchDomainConfig', updateElasticsearchDomainConfig_autoTuneOptions - Specifies Auto-Tune options.

$sel:logPublishingOptions:UpdateElasticsearchDomainConfig', updateElasticsearchDomainConfig_logPublishingOptions - Map of LogType and LogPublishingOption, each containing options to publish a given type of Elasticsearch log.

$sel:advancedSecurityOptions:UpdateElasticsearchDomainConfig', updateElasticsearchDomainConfig_advancedSecurityOptions - Specifies advanced security options.

$sel:elasticsearchClusterConfig:UpdateElasticsearchDomainConfig', updateElasticsearchDomainConfig_elasticsearchClusterConfig - The type and number of instances to instantiate for the domain cluster.

$sel:snapshotOptions:UpdateElasticsearchDomainConfig', updateElasticsearchDomainConfig_snapshotOptions - Option to set the time, in UTC format, for the daily automated snapshot. Default value is 0 hours.

$sel:cognitoOptions:UpdateElasticsearchDomainConfig', updateElasticsearchDomainConfig_cognitoOptions - Options to specify the Cognito user and identity pools for Kibana authentication. For more information, see Amazon Cognito Authentication for Kibana.

$sel:encryptionAtRestOptions:UpdateElasticsearchDomainConfig', updateElasticsearchDomainConfig_encryptionAtRestOptions - Specifies the Encryption At Rest Options.

$sel:vPCOptions:UpdateElasticsearchDomainConfig', updateElasticsearchDomainConfig_vPCOptions - Options to specify the subnets and security groups for VPC endpoint. For more information, see Creating a VPC in VPC Endpoints for Amazon Elasticsearch Service Domains

$sel:domainEndpointOptions:UpdateElasticsearchDomainConfig', updateElasticsearchDomainConfig_domainEndpointOptions - Options to specify configuration that will be applied to the domain endpoint.

$sel:advancedOptions:UpdateElasticsearchDomainConfig', updateElasticsearchDomainConfig_advancedOptions - Modifies the advanced option to allow references to indices in an HTTP request body. Must be false when configuring access to individual sub-resources. By default, the value is true. See Configuration Advanced Options for more information.

$sel:domainName:UpdateElasticsearchDomainConfig', updateElasticsearchDomainConfig_domainName - The name of the Elasticsearch domain that you are updating.

Request Lenses

updateElasticsearchDomainConfig_eBSOptions :: Lens' UpdateElasticsearchDomainConfig (Maybe EBSOptions)

Specify the type and size of the EBS volume that you want to use.

updateElasticsearchDomainConfig_nodeToNodeEncryptionOptions :: Lens' UpdateElasticsearchDomainConfig (Maybe NodeToNodeEncryptionOptions)

Specifies the NodeToNodeEncryptionOptions.

updateElasticsearchDomainConfig_accessPolicies :: Lens' UpdateElasticsearchDomainConfig (Maybe Text)

IAM access policy as a JSON-formatted string.

updateElasticsearchDomainConfig_autoTuneOptions :: Lens' UpdateElasticsearchDomainConfig (Maybe AutoTuneOptions)

Specifies Auto-Tune options.

updateElasticsearchDomainConfig_logPublishingOptions :: Lens' UpdateElasticsearchDomainConfig (Maybe (HashMap LogType LogPublishingOption))

Map of LogType and LogPublishingOption, each containing options to publish a given type of Elasticsearch log.

updateElasticsearchDomainConfig_advancedSecurityOptions :: Lens' UpdateElasticsearchDomainConfig (Maybe AdvancedSecurityOptionsInput)

Specifies advanced security options.

updateElasticsearchDomainConfig_elasticsearchClusterConfig :: Lens' UpdateElasticsearchDomainConfig (Maybe ElasticsearchClusterConfig)

The type and number of instances to instantiate for the domain cluster.

updateElasticsearchDomainConfig_snapshotOptions :: Lens' UpdateElasticsearchDomainConfig (Maybe SnapshotOptions)

Option to set the time, in UTC format, for the daily automated snapshot. Default value is 0 hours.

updateElasticsearchDomainConfig_cognitoOptions :: Lens' UpdateElasticsearchDomainConfig (Maybe CognitoOptions)

Options to specify the Cognito user and identity pools for Kibana authentication. For more information, see Amazon Cognito Authentication for Kibana.

updateElasticsearchDomainConfig_encryptionAtRestOptions :: Lens' UpdateElasticsearchDomainConfig (Maybe EncryptionAtRestOptions)

Specifies the Encryption At Rest Options.

updateElasticsearchDomainConfig_vPCOptions :: Lens' UpdateElasticsearchDomainConfig (Maybe VPCOptions)

Options to specify the subnets and security groups for VPC endpoint. For more information, see Creating a VPC in VPC Endpoints for Amazon Elasticsearch Service Domains

updateElasticsearchDomainConfig_domainEndpointOptions :: Lens' UpdateElasticsearchDomainConfig (Maybe DomainEndpointOptions)

Options to specify configuration that will be applied to the domain endpoint.

updateElasticsearchDomainConfig_advancedOptions :: Lens' UpdateElasticsearchDomainConfig (Maybe (HashMap Text Text))

Modifies the advanced option to allow references to indices in an HTTP request body. Must be false when configuring access to individual sub-resources. By default, the value is true. See Configuration Advanced Options for more information.

updateElasticsearchDomainConfig_domainName :: Lens' UpdateElasticsearchDomainConfig Text

The name of the Elasticsearch domain that you are updating.

Destructuring the Response

data UpdateElasticsearchDomainConfigResponse

The result of an UpdateElasticsearchDomain request. Contains the status of the Elasticsearch domain being updated.

See: newUpdateElasticsearchDomainConfigResponse smart constructor.

Constructors

UpdateElasticsearchDomainConfigResponse'

Fields

httpStatus :: Int

The response's http status code.

domainConfig :: ElasticsearchDomainConfig

The status of the updated Elasticsearch domain.

newUpdateElasticsearchDomainConfigResponse

Arguments

:: Int
$sel:httpStatus:UpdateElasticsearchDomainConfigResponse'

-> ElasticsearchDomainConfig
$sel:domainConfig:UpdateElasticsearchDomainConfigResponse'

-> UpdateElasticsearchDomainConfigResponse

Create a value of UpdateElasticsearchDomainConfigResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:httpStatus:UpdateElasticsearchDomainConfigResponse', updateElasticsearchDomainConfigResponse_httpStatus - The response's http status code.

$sel:domainConfig:UpdateElasticsearchDomainConfigResponse', updateElasticsearchDomainConfigResponse_domainConfig - The status of the updated Elasticsearch domain.

Response Lenses

updateElasticsearchDomainConfigResponse_httpStatus :: Lens' UpdateElasticsearchDomainConfigResponse Int

The response's http status code.

updateElasticsearchDomainConfigResponse_domainConfig :: Lens' UpdateElasticsearchDomainConfigResponse ElasticsearchDomainConfig

The status of the updated Elasticsearch domain.