Amazonka.ElasticTranscoder.Types.Captions

Description

Documentation

data Captions

The captions to be created, if any.

See: newCaptions smart constructor.

Constructors

Captions'

Fields

mergePolicy :: Maybe Text

A policy that determines how Elastic Transcoder handles the existence of multiple captions.

MergeOverride: Elastic Transcoder transcodes both embedded and sidecar captions into outputs. If captions for a language are embedded in the input file and also appear in a sidecar file, Elastic Transcoder uses the sidecar captions and ignores the embedded captions for that language.

MergeRetain: Elastic Transcoder transcodes both embedded and sidecar captions into outputs. If captions for a language are embedded in the input file and also appear in a sidecar file, Elastic Transcoder uses the embedded captions and ignores the sidecar captions for that language. If CaptionSources is empty, Elastic Transcoder omits all sidecar captions from the output files.

Override: Elastic Transcoder transcodes only the sidecar captions that you specify in CaptionSources.

MergePolicy cannot be null.

captionSources :: Maybe [CaptionSource]

Source files for the input sidecar captions used during the transcoding process. To omit all sidecar captions, leave CaptionSources blank.

captionFormats :: Maybe [CaptionFormat]

The array of file formats for the output captions. If you leave this value blank, Elastic Transcoder returns an error.

newCaptions :: Captions

Create a value of Captions with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:mergePolicy:Captions', captions_mergePolicy - A policy that determines how Elastic Transcoder handles the existence of multiple captions.

MergeOverride: Elastic Transcoder transcodes both embedded and sidecar captions into outputs. If captions for a language are embedded in the input file and also appear in a sidecar file, Elastic Transcoder uses the sidecar captions and ignores the embedded captions for that language.

MergeRetain: Elastic Transcoder transcodes both embedded and sidecar captions into outputs. If captions for a language are embedded in the input file and also appear in a sidecar file, Elastic Transcoder uses the embedded captions and ignores the sidecar captions for that language. If CaptionSources is empty, Elastic Transcoder omits all sidecar captions from the output files.

Override: Elastic Transcoder transcodes only the sidecar captions that you specify in CaptionSources.

MergePolicy cannot be null.

$sel:captionSources:Captions', captions_captionSources - Source files for the input sidecar captions used during the transcoding process. To omit all sidecar captions, leave CaptionSources blank.

$sel:captionFormats:Captions', captions_captionFormats - The array of file formats for the output captions. If you leave this value blank, Elastic Transcoder returns an error.

captions_mergePolicy :: Lens' Captions (Maybe Text)

A policy that determines how Elastic Transcoder handles the existence of multiple captions.

MergeOverride: Elastic Transcoder transcodes both embedded and sidecar captions into outputs. If captions for a language are embedded in the input file and also appear in a sidecar file, Elastic Transcoder uses the sidecar captions and ignores the embedded captions for that language.

MergeRetain: Elastic Transcoder transcodes both embedded and sidecar captions into outputs. If captions for a language are embedded in the input file and also appear in a sidecar file, Elastic Transcoder uses the embedded captions and ignores the sidecar captions for that language. If CaptionSources is empty, Elastic Transcoder omits all sidecar captions from the output files.

Override: Elastic Transcoder transcodes only the sidecar captions that you specify in CaptionSources.

MergePolicy cannot be null.

captions_captionSources :: Lens' Captions (Maybe [CaptionSource])

Source files for the input sidecar captions used during the transcoding process. To omit all sidecar captions, leave CaptionSources blank.

captions_captionFormats :: Lens' Captions (Maybe [CaptionFormat])

The array of file formats for the output captions. If you leave this value blank, Elastic Transcoder returns an error.