Amazonka.ElasticTranscoder.Types.Clip

Description

Documentation

data Clip

Settings for one clip in a composition. All jobs in a playlist must have the same clip settings.

See: newClip smart constructor.

Constructors

Clip'

Fields

timeSpan :: Maybe TimeSpan

Settings that determine when a clip begins and how long it lasts.

newClip :: Clip

Create a value of Clip with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:timeSpan:Clip', clip_timeSpan - Settings that determine when a clip begins and how long it lasts.

clip_timeSpan :: Lens' Clip (Maybe TimeSpan)

Settings that determine when a clip begins and how long it lasts.