Amazonka.ElasticTranscoder.Types.Job

Description

Documentation

data Job

A section of the response body that provides information about the job that is created.

See: newJob smart constructor.

Constructors

Job'

Fields

status :: Maybe Text

The status of the job: Submitted, Progressing, Complete, Canceled, or Error.

pipelineId :: Maybe Text

The Id of the pipeline that you want Elastic Transcoder to use for transcoding. The pipeline determines several settings, including the Amazon S3 bucket from which Elastic Transcoder gets the files to transcode and the bucket into which Elastic Transcoder puts the transcoded files.

arn :: Maybe Text

The Amazon Resource Name (ARN) for the job.

inputs :: Maybe [JobInput]

Information about the files that you're transcoding. If you specified multiple files for this job, Elastic Transcoder stitches the files together to make one output.

input :: Maybe JobInput

A section of the request or response body that provides information about the file that is being transcoded.

userMetadata :: Maybe (HashMap Text Text)

User-defined metadata that you want to associate with an Elastic Transcoder job. You specify metadata in key/value pairs, and you can add up to 10 key/value pairs per job. Elastic Transcoder does not guarantee that key/value pairs are returned in the same order in which you specify them.

Metadata keys and values must use characters from the following list:

0-9

A-Z and a-z

Space

The following symbols: _.:/=+-%@

outputs :: Maybe [JobOutput]

Information about the output files. We recommend that you use the Outputs syntax for all jobs, even when you want Elastic Transcoder to transcode a file into only one format. Do not use both the Outputs and Output syntaxes in the same request. You can create a maximum of 30 outputs per job.

If you specify more than one output for a job, Elastic Transcoder creates the files for each output in the order in which you specify them in the job.

output :: Maybe JobOutput

If you specified one output for a job, information about that output. If you specified multiple outputs for a job, the Output object lists information about the first output. This duplicates the information that is listed for the first output in the Outputs object.

Outputs recommended instead.

A section of the request or response body that provides information about the transcoded (target) file.

id :: Maybe Text

The identifier that Elastic Transcoder assigned to the job. You use this value to get settings for the job or to delete the job.

playlists :: Maybe [Playlist]

Outputs in Fragmented MP4 or MPEG-TS format only.

If you specify a preset in PresetId for which the value of Container is fmp4 (Fragmented MP4) or ts (MPEG-TS), Playlists contains information about the master playlists that you want Elastic Transcoder to create.

The maximum number of master playlists in a job is 30.

outputKeyPrefix :: Maybe Text

The value, if any, that you want Elastic Transcoder to prepend to the names of all files that this job creates, including output files, thumbnails, and playlists. We recommend that you add a / or some other delimiter to the end of the OutputKeyPrefix.

timing :: Maybe Timing

Details about the timing of a job.

newJob :: Job

Create a value of Job with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:status:Job', job_status - The status of the job: Submitted, Progressing, Complete, Canceled, or Error.

$sel:pipelineId:Job', job_pipelineId - The Id of the pipeline that you want Elastic Transcoder to use for transcoding. The pipeline determines several settings, including the Amazon S3 bucket from which Elastic Transcoder gets the files to transcode and the bucket into which Elastic Transcoder puts the transcoded files.

$sel:arn:Job', job_arn - The Amazon Resource Name (ARN) for the job.

$sel:inputs:Job', job_inputs - Information about the files that you're transcoding. If you specified multiple files for this job, Elastic Transcoder stitches the files together to make one output.

$sel:input:Job', job_input - A section of the request or response body that provides information about the file that is being transcoded.

$sel:userMetadata:Job', job_userMetadata - User-defined metadata that you want to associate with an Elastic Transcoder job. You specify metadata in key/value pairs, and you can add up to 10 key/value pairs per job. Elastic Transcoder does not guarantee that key/value pairs are returned in the same order in which you specify them.

Metadata keys and values must use characters from the following list:

0-9

A-Z and a-z

Space

The following symbols: _.:/=+-%@

$sel:outputs:Job', job_outputs - Information about the output files. We recommend that you use the Outputs syntax for all jobs, even when you want Elastic Transcoder to transcode a file into only one format. Do not use both the Outputs and Output syntaxes in the same request. You can create a maximum of 30 outputs per job.

If you specify more than one output for a job, Elastic Transcoder creates the files for each output in the order in which you specify them in the job.

$sel:output:Job', job_output - If you specified one output for a job, information about that output. If you specified multiple outputs for a job, the Output object lists information about the first output. This duplicates the information that is listed for the first output in the Outputs object.

Outputs recommended instead.

A section of the request or response body that provides information about the transcoded (target) file.

$sel:id:Job', job_id - The identifier that Elastic Transcoder assigned to the job. You use this value to get settings for the job or to delete the job.

$sel:playlists:Job', job_playlists - Outputs in Fragmented MP4 or MPEG-TS format only.

If you specify a preset in PresetId for which the value of Container is fmp4 (Fragmented MP4) or ts (MPEG-TS), Playlists contains information about the master playlists that you want Elastic Transcoder to create.

The maximum number of master playlists in a job is 30.

$sel:outputKeyPrefix:Job', job_outputKeyPrefix - The value, if any, that you want Elastic Transcoder to prepend to the names of all files that this job creates, including output files, thumbnails, and playlists. We recommend that you add a / or some other delimiter to the end of the OutputKeyPrefix.

$sel:timing:Job', job_timing - Details about the timing of a job.

job_status :: Lens' Job (Maybe Text)

The status of the job: Submitted, Progressing, Complete, Canceled, or Error.

job_pipelineId :: Lens' Job (Maybe Text)

The Id of the pipeline that you want Elastic Transcoder to use for transcoding. The pipeline determines several settings, including the Amazon S3 bucket from which Elastic Transcoder gets the files to transcode and the bucket into which Elastic Transcoder puts the transcoded files.

job_arn :: Lens' Job (Maybe Text)

The Amazon Resource Name (ARN) for the job.

job_inputs :: Lens' Job (Maybe [JobInput])

Information about the files that you're transcoding. If you specified multiple files for this job, Elastic Transcoder stitches the files together to make one output.

job_input :: Lens' Job (Maybe JobInput)

A section of the request or response body that provides information about the file that is being transcoded.

job_userMetadata :: Lens' Job (Maybe (HashMap Text Text))

User-defined metadata that you want to associate with an Elastic Transcoder job. You specify metadata in key/value pairs, and you can add up to 10 key/value pairs per job. Elastic Transcoder does not guarantee that key/value pairs are returned in the same order in which you specify them.

Metadata keys and values must use characters from the following list:

0-9

A-Z and a-z

Space

The following symbols: _.:/=+-%@

job_outputs :: Lens' Job (Maybe [JobOutput])

Information about the output files. We recommend that you use the Outputs syntax for all jobs, even when you want Elastic Transcoder to transcode a file into only one format. Do not use both the Outputs and Output syntaxes in the same request. You can create a maximum of 30 outputs per job.

If you specify more than one output for a job, Elastic Transcoder creates the files for each output in the order in which you specify them in the job.

job_output :: Lens' Job (Maybe JobOutput)

If you specified one output for a job, information about that output. If you specified multiple outputs for a job, the Output object lists information about the first output. This duplicates the information that is listed for the first output in the Outputs object.

Outputs recommended instead.

A section of the request or response body that provides information about the transcoded (target) file.

job_id :: Lens' Job (Maybe Text)

The identifier that Elastic Transcoder assigned to the job. You use this value to get settings for the job or to delete the job.

job_playlists :: Lens' Job (Maybe [Playlist])

Outputs in Fragmented MP4 or MPEG-TS format only.

If you specify a preset in PresetId for which the value of Container is fmp4 (Fragmented MP4) or ts (MPEG-TS), Playlists contains information about the master playlists that you want Elastic Transcoder to create.

The maximum number of master playlists in a job is 30.

job_outputKeyPrefix :: Lens' Job (Maybe Text)

The value, if any, that you want Elastic Transcoder to prepend to the names of all files that this job creates, including output files, thumbnails, and playlists. We recommend that you add a / or some other delimiter to the end of the OutputKeyPrefix.

job_timing :: Lens' Job (Maybe Timing)

Details about the timing of a job.