Amazonka.ElasticTranscoder.UpdatePipelineNotifications

Description

With the UpdatePipelineNotifications operation, you can update Amazon Simple Notification Service (Amazon SNS) notifications for a pipeline.

When you update notifications for a pipeline, Elastic Transcoder returns the values that you specified in the request.

Creating a Request

data UpdatePipelineNotifications

The UpdatePipelineNotificationsRequest structure.

See: newUpdatePipelineNotifications smart constructor.

Constructors

UpdatePipelineNotifications'

Fields

id :: Text

The identifier of the pipeline for which you want to change notification settings.

notifications :: Notifications

The topic ARN for the Amazon Simple Notification Service (Amazon SNS) topic that you want to notify to report job status.

To receive notifications, you must also subscribe to the new topic in the Amazon SNS console.

Progressing: The topic ARN for the Amazon Simple Notification Service (Amazon SNS) topic that you want to notify when Elastic Transcoder has started to process jobs that are added to this pipeline. This is the ARN that Amazon SNS returned when you created the topic.

Complete: The topic ARN for the Amazon SNS topic that you want to notify when Elastic Transcoder has finished processing a job. This is the ARN that Amazon SNS returned when you created the topic.

Warning: The topic ARN for the Amazon SNS topic that you want to notify when Elastic Transcoder encounters a warning condition. This is the ARN that Amazon SNS returned when you created the topic.

Error: The topic ARN for the Amazon SNS topic that you want to notify when Elastic Transcoder encounters an error condition. This is the ARN that Amazon SNS returned when you created the topic.

newUpdatePipelineNotifications

Arguments

:: Text
$sel:id:UpdatePipelineNotifications'

-> Notifications
$sel:notifications:UpdatePipelineNotifications'

-> UpdatePipelineNotifications

Create a value of UpdatePipelineNotifications with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:id:UpdatePipelineNotifications', updatePipelineNotifications_id - The identifier of the pipeline for which you want to change notification settings.

$sel:notifications:UpdatePipelineNotifications', updatePipelineNotifications_notifications - The topic ARN for the Amazon Simple Notification Service (Amazon SNS) topic that you want to notify to report job status.

To receive notifications, you must also subscribe to the new topic in the Amazon SNS console.

Progressing: The topic ARN for the Amazon Simple Notification Service (Amazon SNS) topic that you want to notify when Elastic Transcoder has started to process jobs that are added to this pipeline. This is the ARN that Amazon SNS returned when you created the topic.

Complete: The topic ARN for the Amazon SNS topic that you want to notify when Elastic Transcoder has finished processing a job. This is the ARN that Amazon SNS returned when you created the topic.

Warning: The topic ARN for the Amazon SNS topic that you want to notify when Elastic Transcoder encounters a warning condition. This is the ARN that Amazon SNS returned when you created the topic.

Error: The topic ARN for the Amazon SNS topic that you want to notify when Elastic Transcoder encounters an error condition. This is the ARN that Amazon SNS returned when you created the topic.

Request Lenses

updatePipelineNotifications_id :: Lens' UpdatePipelineNotifications Text

The identifier of the pipeline for which you want to change notification settings.

updatePipelineNotifications_notifications :: Lens' UpdatePipelineNotifications Notifications

The topic ARN for the Amazon Simple Notification Service (Amazon SNS) topic that you want to notify to report job status.

To receive notifications, you must also subscribe to the new topic in the Amazon SNS console.

Progressing: The topic ARN for the Amazon Simple Notification Service (Amazon SNS) topic that you want to notify when Elastic Transcoder has started to process jobs that are added to this pipeline. This is the ARN that Amazon SNS returned when you created the topic.

Complete: The topic ARN for the Amazon SNS topic that you want to notify when Elastic Transcoder has finished processing a job. This is the ARN that Amazon SNS returned when you created the topic.

Warning: The topic ARN for the Amazon SNS topic that you want to notify when Elastic Transcoder encounters a warning condition. This is the ARN that Amazon SNS returned when you created the topic.

Error: The topic ARN for the Amazon SNS topic that you want to notify when Elastic Transcoder encounters an error condition. This is the ARN that Amazon SNS returned when you created the topic.

Destructuring the Response

data UpdatePipelineNotificationsResponse

The UpdatePipelineNotificationsResponse structure.

See: newUpdatePipelineNotificationsResponse smart constructor.

Constructors

UpdatePipelineNotificationsResponse'

Fields

pipeline :: Maybe Pipeline

A section of the response body that provides information about the pipeline associated with this notification.

httpStatus :: Int

The response's http status code.

newUpdatePipelineNotificationsResponse

Arguments

:: Int
$sel:httpStatus:UpdatePipelineNotificationsResponse'

-> UpdatePipelineNotificationsResponse

Create a value of UpdatePipelineNotificationsResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:pipeline:UpdatePipelineNotificationsResponse', updatePipelineNotificationsResponse_pipeline - A section of the response body that provides information about the pipeline associated with this notification.

$sel:httpStatus:UpdatePipelineNotificationsResponse', updatePipelineNotificationsResponse_httpStatus - The response's http status code.

Response Lenses

updatePipelineNotificationsResponse_pipeline :: Lens' UpdatePipelineNotificationsResponse (Maybe Pipeline)

A section of the response body that provides information about the pipeline associated with this notification.

updatePipelineNotificationsResponse_httpStatus :: Lens' UpdatePipelineNotificationsResponse Int

The response's http status code.