Amazonka.Error

Description

Documentation

_MatchServiceError :: AsError a => Service -> ErrorCode -> Getting (First ServiceError) a ServiceError

Provides a generalised prism for catching a specific service error identified by the opaque service abbreviation and error code.

This can be used if the generated error prisms provided by Amazonka.ServiceName.Types do not cover all the thrown error codes. For example to define a new error prism:

{-# LANGUAGE OverloadedStrings #-}

import Amazonka.S3 (ServiceError, s3)

_NoSuchBucketPolicy :: AsError a => Getting (First ServiceError) a ServiceError
_NoSuchBucketPolicy = _MatchServiceError s3 "NoSuchBucketPolicy"

With example usage being:

>>> import Control.Exception.Lens (trying)
>>> :t trying _NoSuchBucketPolicy
MonadCatch m => m a -> m (Either ServiceError a)

statusSuccess :: Status -> Bool

_HttpStatus :: AsError a => Getting (First Status) a Status

hasService :: (Applicative f, Choice p) => Service -> Optic' p f ServiceError ServiceError

hasStatus :: (Applicative f, Choice p) => Int -> Optic' p f ServiceError ServiceError

hasCode :: (Applicative f, Choice p) => ErrorCode -> Optic' p f ServiceError ServiceError

serviceError :: Abbrev -> Status -> [Header] -> Maybe ErrorCode -> Maybe ErrorMessage -> Maybe RequestId -> ServiceError

getRequestId :: [Header] -> Maybe RequestId

getErrorCode :: Status -> [Header] -> ErrorCode

parseJSONError :: Abbrev -> Status -> [Header] -> ByteStringLazy -> Error

parseXMLError :: Abbrev -> Status -> [Header] -> ByteStringLazy -> Error

parseRESTError :: Abbrev -> Status -> [Header] -> a -> Error

decodeError :: Abbrev -> Status -> [Header] -> ByteStringLazy -> Either String ServiceError -> Error