Amazonka.FIS

Description

Derived from API version 2020-12-01 of the AWS service descriptions, licensed under Apache 2.0.

AWS Fault Injection Simulator is a managed service that enables you to perform fault injection experiments on your AWS workloads. For more information, see the AWS Fault Injection Simulator User Guide.

Service Configuration

defaultService :: Service

API version 2020-12-01 of the Amazon Fault Injection Simulator SDK configuration.

Errors

Error matchers are designed for use with the functions provided by Control.Exception.Lens. This allows catching (and rethrowing) service specific errors returned by FIS.

ValidationException

_ValidationException :: AsError a => Getting (First ServiceError) a ServiceError

The specified input is not valid, or fails to satisfy the constraints for the request.

ConflictException

_ConflictException :: AsError a => Getting (First ServiceError) a ServiceError

The request could not be processed because of a conflict.

ServiceQuotaExceededException

_ServiceQuotaExceededException :: AsError a => Getting (First ServiceError) a ServiceError

You have exceeded your service quota.

ResourceNotFoundException

_ResourceNotFoundException :: AsError a => Getting (First ServiceError) a ServiceError

The specified resource cannot be found.

Waiters

Waiters poll by repeatedly sending a request until some remote success condition configured by the Wait specification is fulfilled. The Wait specification determines how many attempts should be made, in addition to delay and retry strategies.

Operations

Some AWS operations return results that are incomplete and require subsequent requests in order to obtain the entire result set. The process of sending subsequent requests to continue where a previous request left off is called pagination. For example, the ListObjects operation of Amazon S3 returns up to 1000 objects at a time, and you must send subsequent requests with the appropriate Marker in order to retrieve the next page of results.

Operations that have an AWSPager instance can transparently perform subsequent requests, correctly setting Markers and other request facets to iterate through the entire result set of a truncated API operation. Operations which support this have an additional note in the documentation.

Many operations have the ability to filter results on the server side. See the individual operation parameters for details.

GetExperimentTemplate

data GetExperimentTemplate

See: newGetExperimentTemplate smart constructor.

Constructors

GetExperimentTemplate' Text

newGetExperimentTemplate

Arguments

:: Text
$sel:id:GetExperimentTemplate'

-> GetExperimentTemplate

Create a value of GetExperimentTemplate with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:id:GetExperimentTemplate', getExperimentTemplate_id - The ID of the experiment template.

data GetExperimentTemplateResponse

See: newGetExperimentTemplateResponse smart constructor.

Constructors

GetExperimentTemplateResponse' (Maybe ExperimentTemplate) Int

newGetExperimentTemplateResponse

Arguments

:: Int
$sel:httpStatus:GetExperimentTemplateResponse'

-> GetExperimentTemplateResponse

Create a value of GetExperimentTemplateResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:experimentTemplate:GetExperimentTemplateResponse', getExperimentTemplateResponse_experimentTemplate - Information about the experiment template.

$sel:httpStatus:GetExperimentTemplateResponse', getExperimentTemplateResponse_httpStatus - The response's http status code.

ListActions

data ListActions

See: newListActions smart constructor.

Constructors

ListActions' (Maybe Text) (Maybe Natural)

newListActions :: ListActions

Create a value of ListActions with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:nextToken:ListActions', listActions_nextToken - The token for the next page of results.

$sel:maxResults:ListActions', listActions_maxResults - The maximum number of results to return with a single call. To retrieve the remaining results, make another call with the returned nextToken value.

data ListActionsResponse

See: newListActionsResponse smart constructor.

Constructors

ListActionsResponse' (Maybe [ActionSummary]) (Maybe Text) Int

newListActionsResponse

Arguments

:: Int
$sel:httpStatus:ListActionsResponse'

-> ListActionsResponse

Create a value of ListActionsResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:actions:ListActionsResponse', listActionsResponse_actions - The actions.

$sel:nextToken:ListActions', listActionsResponse_nextToken - The token to use to retrieve the next page of results. This value is null when there are no more results to return.

$sel:httpStatus:ListActionsResponse', listActionsResponse_httpStatus - The response's http status code.

ListTagsForResource

data ListTagsForResource

See: newListTagsForResource smart constructor.

Constructors

ListTagsForResource' Text

newListTagsForResource

Arguments

:: Text
$sel:resourceArn:ListTagsForResource'

-> ListTagsForResource

Create a value of ListTagsForResource with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:resourceArn:ListTagsForResource', listTagsForResource_resourceArn - The Amazon Resource Name (ARN) of the resource.

data ListTagsForResourceResponse

See: newListTagsForResourceResponse smart constructor.

Constructors

ListTagsForResourceResponse' (Maybe (HashMap Text Text)) Int

newListTagsForResourceResponse

Arguments

:: Int
$sel:httpStatus:ListTagsForResourceResponse'

-> ListTagsForResourceResponse

Create a value of ListTagsForResourceResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:tags:ListTagsForResourceResponse', listTagsForResourceResponse_tags - The tags for the resource.

$sel:httpStatus:ListTagsForResourceResponse', listTagsForResourceResponse_httpStatus - The response's http status code.

CreateExperimentTemplate

data CreateExperimentTemplate

See: newCreateExperimentTemplate smart constructor.

Constructors

CreateExperimentTemplate' (Maybe (HashMap Text CreateExperimentTemplateTargetInput)) (Maybe (HashMap Text Text)) Text Text [CreateExperimentTemplateStopConditionInput] (HashMap Text CreateExperimentTemplateActionInput) Text

newCreateExperimentTemplate

Arguments

:: Text
$sel:clientToken:CreateExperimentTemplate'

-> Text
$sel:description:CreateExperimentTemplate'

-> Text
$sel:roleArn:CreateExperimentTemplate'

-> CreateExperimentTemplate

Create a value of CreateExperimentTemplate with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:targets:CreateExperimentTemplate', createExperimentTemplate_targets - The targets for the experiment.

$sel:tags:CreateExperimentTemplate', createExperimentTemplate_tags - The tags to apply to the experiment template.

$sel:clientToken:CreateExperimentTemplate', createExperimentTemplate_clientToken - Unique, case-sensitive identifier that you provide to ensure the idempotency of the request.

$sel:description:CreateExperimentTemplate', createExperimentTemplate_description - A description for the experiment template. Can contain up to 64 letters (A-Z and a-z).

$sel:stopConditions:CreateExperimentTemplate', createExperimentTemplate_stopConditions - The stop conditions.

$sel:actions:CreateExperimentTemplate', createExperimentTemplate_actions - The actions for the experiment.

$sel:roleArn:CreateExperimentTemplate', createExperimentTemplate_roleArn - The Amazon Resource Name (ARN) of an IAM role that grants the AWS FIS service permission to perform service actions on your behalf.

data CreateExperimentTemplateResponse

See: newCreateExperimentTemplateResponse smart constructor.

Constructors

CreateExperimentTemplateResponse' (Maybe ExperimentTemplate) Int

newCreateExperimentTemplateResponse

Arguments

:: Int
$sel:httpStatus:CreateExperimentTemplateResponse'

-> CreateExperimentTemplateResponse

Create a value of CreateExperimentTemplateResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:experimentTemplate:CreateExperimentTemplateResponse', createExperimentTemplateResponse_experimentTemplate - Information about the experiment template.

$sel:httpStatus:CreateExperimentTemplateResponse', createExperimentTemplateResponse_httpStatus - The response's http status code.

ListExperiments

data ListExperiments

See: newListExperiments smart constructor.

Constructors

ListExperiments' (Maybe Text) (Maybe Natural)

newListExperiments :: ListExperiments

Create a value of ListExperiments with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:nextToken:ListExperiments', listExperiments_nextToken - The token for the next page of results.

$sel:maxResults:ListExperiments', listExperiments_maxResults - The maximum number of results to return with a single call. To retrieve the remaining results, make another call with the returned nextToken value.

data ListExperimentsResponse

See: newListExperimentsResponse smart constructor.

Constructors

ListExperimentsResponse' (Maybe [ExperimentSummary]) (Maybe Text) Int

newListExperimentsResponse

Arguments

:: Int
$sel:httpStatus:ListExperimentsResponse'

-> ListExperimentsResponse

Create a value of ListExperimentsResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:experiments:ListExperimentsResponse', listExperimentsResponse_experiments - The experiments.

$sel:nextToken:ListExperiments', listExperimentsResponse_nextToken - The token to use to retrieve the next page of results. This value is null when there are no more results to return.

$sel:httpStatus:ListExperimentsResponse', listExperimentsResponse_httpStatus - The response's http status code.

UpdateExperimentTemplate

data UpdateExperimentTemplate

See: newUpdateExperimentTemplate smart constructor.

Constructors

UpdateExperimentTemplate' (Maybe (HashMap Text UpdateExperimentTemplateActionInputItem)) (Maybe [UpdateExperimentTemplateStopConditionInput]) (Maybe (HashMap Text UpdateExperimentTemplateTargetInput)) (Maybe Text) (Maybe Text) Text

newUpdateExperimentTemplate

Arguments

:: Text
$sel:id:UpdateExperimentTemplate'

-> UpdateExperimentTemplate

Create a value of UpdateExperimentTemplate with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:actions:UpdateExperimentTemplate', updateExperimentTemplate_actions - The actions for the experiment.

$sel:stopConditions:UpdateExperimentTemplate', updateExperimentTemplate_stopConditions - The stop conditions for the experiment.

$sel:targets:UpdateExperimentTemplate', updateExperimentTemplate_targets - The targets for the experiment.

$sel:description:UpdateExperimentTemplate', updateExperimentTemplate_description - A description for the template.

$sel:roleArn:UpdateExperimentTemplate', updateExperimentTemplate_roleArn - The Amazon Resource Name (ARN) of an IAM role that grants the AWS FIS service permission to perform service actions on your behalf.

$sel:id:UpdateExperimentTemplate', updateExperimentTemplate_id - The ID of the experiment template.

data UpdateExperimentTemplateResponse

See: newUpdateExperimentTemplateResponse smart constructor.

Constructors

UpdateExperimentTemplateResponse' (Maybe ExperimentTemplate) Int

newUpdateExperimentTemplateResponse

Arguments

:: Int
$sel:httpStatus:UpdateExperimentTemplateResponse'

-> UpdateExperimentTemplateResponse

Create a value of UpdateExperimentTemplateResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:experimentTemplate:UpdateExperimentTemplateResponse', updateExperimentTemplateResponse_experimentTemplate - Information about the experiment template.

$sel:httpStatus:UpdateExperimentTemplateResponse', updateExperimentTemplateResponse_httpStatus - The response's http status code.

GetAction

data GetAction

See: newGetAction smart constructor.

Constructors

GetAction' Text

newGetAction

Arguments

:: Text
$sel:id:GetAction'

-> GetAction

Create a value of GetAction with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:id:GetAction', getAction_id - The ID of the action.

data GetActionResponse

See: newGetActionResponse smart constructor.

Constructors

GetActionResponse' (Maybe Action) Int

newGetActionResponse

Arguments

:: Int
$sel:httpStatus:GetActionResponse'

-> GetActionResponse

Create a value of GetActionResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:action:GetActionResponse', getActionResponse_action - Information about the action.

$sel:httpStatus:GetActionResponse', getActionResponse_httpStatus - The response's http status code.

DeleteExperimentTemplate

data DeleteExperimentTemplate

See: newDeleteExperimentTemplate smart constructor.

Constructors

DeleteExperimentTemplate' Text

newDeleteExperimentTemplate

Arguments

:: Text
$sel:id:DeleteExperimentTemplate'

-> DeleteExperimentTemplate

Create a value of DeleteExperimentTemplate with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:id:DeleteExperimentTemplate', deleteExperimentTemplate_id - The ID of the experiment template.

data DeleteExperimentTemplateResponse

See: newDeleteExperimentTemplateResponse smart constructor.

Constructors

DeleteExperimentTemplateResponse' (Maybe ExperimentTemplate) Int

newDeleteExperimentTemplateResponse

Arguments

:: Int
$sel:httpStatus:DeleteExperimentTemplateResponse'

-> DeleteExperimentTemplateResponse

Create a value of DeleteExperimentTemplateResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:experimentTemplate:DeleteExperimentTemplateResponse', deleteExperimentTemplateResponse_experimentTemplate - Information about the experiment template.

$sel:httpStatus:DeleteExperimentTemplateResponse', deleteExperimentTemplateResponse_httpStatus - The response's http status code.

StartExperiment

data StartExperiment

See: newStartExperiment smart constructor.

Constructors

StartExperiment' (Maybe (HashMap Text Text)) Text Text

newStartExperiment

Arguments

:: Text
$sel:clientToken:StartExperiment'

-> Text
$sel:experimentTemplateId:StartExperiment'

-> StartExperiment

Create a value of StartExperiment with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:tags:StartExperiment', startExperiment_tags - The tags to apply to the experiment.

$sel:clientToken:StartExperiment', startExperiment_clientToken - Unique, case-sensitive identifier that you provide to ensure the idempotency of the request.

$sel:experimentTemplateId:StartExperiment', startExperiment_experimentTemplateId - The ID of the experiment template.

data StartExperimentResponse

See: newStartExperimentResponse smart constructor.

Constructors

StartExperimentResponse' (Maybe Experiment) Int

newStartExperimentResponse

Arguments

:: Int
$sel:httpStatus:StartExperimentResponse'

-> StartExperimentResponse

Create a value of StartExperimentResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:experiment:StartExperimentResponse', startExperimentResponse_experiment - Information about the experiment.

$sel:httpStatus:StartExperimentResponse', startExperimentResponse_httpStatus - The response's http status code.

GetExperiment

data GetExperiment

See: newGetExperiment smart constructor.

Constructors

GetExperiment' Text

newGetExperiment

Arguments

:: Text
$sel:id:GetExperiment'

-> GetExperiment

Create a value of GetExperiment with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:id:GetExperiment', getExperiment_id - The ID of the experiment.

data GetExperimentResponse

See: newGetExperimentResponse smart constructor.

Constructors

GetExperimentResponse' (Maybe Experiment) Int

newGetExperimentResponse

Arguments

:: Int
$sel:httpStatus:GetExperimentResponse'

-> GetExperimentResponse

Create a value of GetExperimentResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:experiment:GetExperimentResponse', getExperimentResponse_experiment - Information about the experiment.

$sel:httpStatus:GetExperimentResponse', getExperimentResponse_httpStatus - The response's http status code.

ListExperimentTemplates

data ListExperimentTemplates

See: newListExperimentTemplates smart constructor.

Constructors

ListExperimentTemplates' (Maybe Text) (Maybe Natural)

newListExperimentTemplates :: ListExperimentTemplates

Create a value of ListExperimentTemplates with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:nextToken:ListExperimentTemplates', listExperimentTemplates_nextToken - The token for the next page of results.

$sel:maxResults:ListExperimentTemplates', listExperimentTemplates_maxResults - The maximum number of results to return with a single call. To retrieve the remaining results, make another call with the returned nextToken value.

data ListExperimentTemplatesResponse

See: newListExperimentTemplatesResponse smart constructor.

Constructors

ListExperimentTemplatesResponse' (Maybe Text) (Maybe [ExperimentTemplateSummary]) Int

newListExperimentTemplatesResponse

Arguments

:: Int
$sel:httpStatus:ListExperimentTemplatesResponse'

-> ListExperimentTemplatesResponse

Create a value of ListExperimentTemplatesResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:nextToken:ListExperimentTemplates', listExperimentTemplatesResponse_nextToken - The token to use to retrieve the next page of results. This value is null when there are no more results to return.

$sel:experimentTemplates:ListExperimentTemplatesResponse', listExperimentTemplatesResponse_experimentTemplates - The experiment templates.

$sel:httpStatus:ListExperimentTemplatesResponse', listExperimentTemplatesResponse_httpStatus - The response's http status code.

TagResource

data TagResource

See: newTagResource smart constructor.

Constructors

TagResource' Text (HashMap Text Text)

newTagResource

Arguments

:: Text
$sel:resourceArn:TagResource'

-> TagResource

Create a value of TagResource with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:resourceArn:TagResource', tagResource_resourceArn - The Amazon Resource Name (ARN) of the resource.

$sel:tags:TagResource', tagResource_tags - The tags for the resource.

data TagResourceResponse

See: newTagResourceResponse smart constructor.

Constructors

TagResourceResponse' Int

newTagResourceResponse

Arguments

:: Int
$sel:httpStatus:TagResourceResponse'

-> TagResourceResponse

Create a value of TagResourceResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:httpStatus:TagResourceResponse', tagResourceResponse_httpStatus - The response's http status code.

UntagResource

data UntagResource

See: newUntagResource smart constructor.

Constructors

UntagResource' (Maybe [Text]) Text

newUntagResource

Arguments

:: Text
$sel:resourceArn:UntagResource'

-> UntagResource

Create a value of UntagResource with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:tagKeys:UntagResource', untagResource_tagKeys - The tag keys to remove.

$sel:resourceArn:UntagResource', untagResource_resourceArn - The Amazon Resource Name (ARN) of the resource.

data UntagResourceResponse

See: newUntagResourceResponse smart constructor.

Constructors

UntagResourceResponse' Int

newUntagResourceResponse

Arguments

:: Int
$sel:httpStatus:UntagResourceResponse'

-> UntagResourceResponse

Create a value of UntagResourceResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:httpStatus:UntagResourceResponse', untagResourceResponse_httpStatus - The response's http status code.

StopExperiment

data StopExperiment

See: newStopExperiment smart constructor.

Constructors

StopExperiment' Text

newStopExperiment

Arguments

:: Text
$sel:id:StopExperiment'

-> StopExperiment

Create a value of StopExperiment with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:id:StopExperiment', stopExperiment_id - The ID of the experiment.

data StopExperimentResponse

See: newStopExperimentResponse smart constructor.

Constructors

StopExperimentResponse' (Maybe Experiment) Int

newStopExperimentResponse

Arguments

:: Int
$sel:httpStatus:StopExperimentResponse'

-> StopExperimentResponse

Create a value of StopExperimentResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:experiment:StopExperimentResponse', stopExperimentResponse_experiment - Information about the experiment.

$sel:httpStatus:StopExperimentResponse', stopExperimentResponse_httpStatus - The response's http status code.

Types

ExperimentActionStatus

newtype ExperimentActionStatus

Constructors

ExperimentActionStatus'

Fields

fromExperimentActionStatus :: Text

Bundled Patterns

pattern ExperimentActionStatus_Cancelled :: ExperimentActionStatus

pattern ExperimentActionStatus_Completed :: ExperimentActionStatus

pattern ExperimentActionStatus_Failed :: ExperimentActionStatus

pattern ExperimentActionStatus_Initiating :: ExperimentActionStatus

pattern ExperimentActionStatus_Pending :: ExperimentActionStatus

pattern ExperimentActionStatus_Running :: ExperimentActionStatus

pattern ExperimentActionStatus_Stopped :: ExperimentActionStatus

pattern ExperimentActionStatus_Stopping :: ExperimentActionStatus

ExperimentStatus

newtype ExperimentStatus

Constructors

ExperimentStatus'

Fields

fromExperimentStatus :: Text

Bundled Patterns

pattern ExperimentStatus_Completed :: ExperimentStatus

pattern ExperimentStatus_Failed :: ExperimentStatus

pattern ExperimentStatus_Initiating :: ExperimentStatus

pattern ExperimentStatus_Pending :: ExperimentStatus

pattern ExperimentStatus_Running :: ExperimentStatus

pattern ExperimentStatus_Stopped :: ExperimentStatus

pattern ExperimentStatus_Stopping :: ExperimentStatus

Action

data Action

Describes an action. For more information, see AWS FIS actions in the AWS Fault Injection Simulator User Guide.

See: newAction smart constructor.

Constructors

Action' (Maybe (HashMap Text ActionParameter)) (Maybe (HashMap Text ActionTarget)) (Maybe Text) (Maybe Text) (Maybe (HashMap Text Text))

newAction :: Action

Create a value of Action with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:parameters:Action', action_parameters - The action parameters, if applicable.

$sel:targets:Action', action_targets - The supported targets for the action.

$sel:id:Action', action_id - The ID of the action.

$sel:description:Action', action_description - The description for the action.

$sel:tags:Action', action_tags - The tags for the action.

ActionParameter

data ActionParameter

Describes a parameter for an action.

See: newActionParameter smart constructor.

Constructors

ActionParameter' (Maybe Bool) (Maybe Text)

newActionParameter :: ActionParameter

Create a value of ActionParameter with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:required:ActionParameter', actionParameter_required - Indicates whether the parameter is required.

$sel:description:ActionParameter', actionParameter_description - The parameter description.

ActionSummary

data ActionSummary

Provides a summary of an action.

See: newActionSummary smart constructor.

Constructors

ActionSummary' (Maybe (HashMap Text ActionTarget)) (Maybe Text) (Maybe Text) (Maybe (HashMap Text Text))

newActionSummary :: ActionSummary

Create a value of ActionSummary with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:targets:ActionSummary', actionSummary_targets - The targets for the action.

$sel:id:ActionSummary', actionSummary_id - The ID of the action.

$sel:description:ActionSummary', actionSummary_description - The description for the action.

$sel:tags:ActionSummary', actionSummary_tags - The tags for the action.

ActionTarget

data ActionTarget

Describes a target for an action.

See: newActionTarget smart constructor.

Constructors

ActionTarget' (Maybe Text)

newActionTarget :: ActionTarget

Create a value of ActionTarget with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:resourceType:ActionTarget', actionTarget_resourceType - The resource type of the target.

CreateExperimentTemplateActionInput

data CreateExperimentTemplateActionInput

Specifies an action for an experiment template.

See: newCreateExperimentTemplateActionInput smart constructor.

Constructors

CreateExperimentTemplateActionInput' (Maybe [Text]) (Maybe (HashMap Text Text)) (Maybe (HashMap Text Text)) (Maybe Text) Text

newCreateExperimentTemplateActionInput

Arguments

:: Text
$sel:actionId:CreateExperimentTemplateActionInput'

-> CreateExperimentTemplateActionInput

Create a value of CreateExperimentTemplateActionInput with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:startAfter:CreateExperimentTemplateActionInput', createExperimentTemplateActionInput_startAfter - The name of the action that must be completed before the current action starts. Omit this parameter to run the action at the start of the experiment.

$sel:parameters:CreateExperimentTemplateActionInput', createExperimentTemplateActionInput_parameters - The parameters for the action, if applicable.

$sel:targets:CreateExperimentTemplateActionInput', createExperimentTemplateActionInput_targets - The targets for the action.

$sel:description:CreateExperimentTemplateActionInput', createExperimentTemplateActionInput_description - A description for the action.

$sel:actionId:CreateExperimentTemplateActionInput', createExperimentTemplateActionInput_actionId - The ID of the action.

CreateExperimentTemplateStopConditionInput

data CreateExperimentTemplateStopConditionInput

Specifies a stop condition for an experiment template.

See: newCreateExperimentTemplateStopConditionInput smart constructor.

Constructors

CreateExperimentTemplateStopConditionInput' (Maybe Text) Text

newCreateExperimentTemplateStopConditionInput

Arguments

:: Text
$sel:source:CreateExperimentTemplateStopConditionInput'

-> CreateExperimentTemplateStopConditionInput

Create a value of CreateExperimentTemplateStopConditionInput with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:value:CreateExperimentTemplateStopConditionInput', createExperimentTemplateStopConditionInput_value - The Amazon Resource Name (ARN) of the CloudWatch alarm. This is required if the source is a CloudWatch alarm.

$sel:source:CreateExperimentTemplateStopConditionInput', createExperimentTemplateStopConditionInput_source - The source for the stop condition. Specify aws:cloudwatch:alarm if the stop condition is defined by a CloudWatch alarm. Specify none if there is no stop condition.

CreateExperimentTemplateTargetInput

data CreateExperimentTemplateTargetInput

Specifies a target for an experiment. You must specify at least one Amazon Resource Name (ARN) or at least one resource tag. You cannot specify both ARNs and tags.

See: newCreateExperimentTemplateTargetInput smart constructor.

Constructors

CreateExperimentTemplateTargetInput' (Maybe (HashMap Text Text)) (Maybe [ExperimentTemplateTargetInputFilter]) (Maybe [Text]) Text Text

newCreateExperimentTemplateTargetInput

Arguments

:: Text
$sel:resourceType:CreateExperimentTemplateTargetInput'

-> Text
$sel:selectionMode:CreateExperimentTemplateTargetInput'

-> CreateExperimentTemplateTargetInput

Create a value of CreateExperimentTemplateTargetInput with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:resourceTags:CreateExperimentTemplateTargetInput', createExperimentTemplateTargetInput_resourceTags - The tags for the target resources.

$sel:filters:CreateExperimentTemplateTargetInput', createExperimentTemplateTargetInput_filters - The filters to apply to identify target resources using specific attributes.

$sel:resourceArns:CreateExperimentTemplateTargetInput', createExperimentTemplateTargetInput_resourceArns - The Amazon Resource Names (ARNs) of the resources.

$sel:resourceType:CreateExperimentTemplateTargetInput', createExperimentTemplateTargetInput_resourceType - The AWS resource type. The resource type must be supported for the specified action.

$sel:selectionMode:CreateExperimentTemplateTargetInput', createExperimentTemplateTargetInput_selectionMode - Scopes the identified resources to a specific count of the resources at random, or a percentage of the resources. All identified resources are included in the target.

ALL - Run the action on all identified targets. This is the default.

COUNT(n) - Run the action on the specified number of targets, chosen from the identified targets at random. For example, COUNT(1) selects one of the targets.

PERCENT(n) - Run the action on the specified percentage of targets, chosen from the identified targets at random. For example, PERCENT(25) selects 25% of the targets.

Experiment

data Experiment

Describes an experiment.

See: newExperiment smart constructor.

Constructors

Experiment' (Maybe POSIX) (Maybe Text) (Maybe ExperimentState) (Maybe (HashMap Text ExperimentAction)) (Maybe POSIX) (Maybe [ExperimentStopCondition]) (Maybe POSIX) (Maybe (HashMap Text ExperimentTarget)) (Maybe Text) (Maybe (HashMap Text Text)) (Maybe Text)

newExperiment :: Experiment

Create a value of Experiment with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:creationTime:Experiment', experiment_creationTime - The time the experiment was created.

$sel:experimentTemplateId:Experiment', experiment_experimentTemplateId - The ID of the experiment template.

$sel:state:Experiment', experiment_state - The state of the experiment.

$sel:actions:Experiment', experiment_actions - The actions for the experiment.

$sel:startTime:Experiment', experiment_startTime - The time that the experiment was started.

$sel:stopConditions:Experiment', experiment_stopConditions - The stop conditions for the experiment.

$sel:endTime:Experiment', experiment_endTime - The time that the experiment ended.

$sel:targets:Experiment', experiment_targets - The targets for the experiment.

$sel:id:Experiment', experiment_id - The ID of the experiment.

$sel:tags:Experiment', experiment_tags - The tags for the experiment.

$sel:roleArn:Experiment', experiment_roleArn - The Amazon Resource Name (ARN) of an IAM role that grants the AWS FIS service permission to perform service actions on your behalf.

ExperimentAction

data ExperimentAction

Describes the action for an experiment.

See: newExperimentAction smart constructor.

Constructors

ExperimentAction' (Maybe [Text]) (Maybe ExperimentActionState) (Maybe Text) (Maybe (HashMap Text Text)) (Maybe (HashMap Text Text)) (Maybe Text)

newExperimentAction :: ExperimentAction

Create a value of ExperimentAction with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:startAfter:ExperimentAction', experimentAction_startAfter - The name of the action that must be completed before this action starts.

$sel:state:ExperimentAction', experimentAction_state - The state of the action.

$sel:actionId:ExperimentAction', experimentAction_actionId - The ID of the action.

$sel:parameters:ExperimentAction', experimentAction_parameters - The parameters for the action.

$sel:targets:ExperimentAction', experimentAction_targets - The targets for the action.

$sel:description:ExperimentAction', experimentAction_description - The description for the action.

ExperimentActionState

data ExperimentActionState

Describes the state of an action.

See: newExperimentActionState smart constructor.

Constructors

ExperimentActionState' (Maybe ExperimentActionStatus) (Maybe Text)

newExperimentActionState :: ExperimentActionState

Create a value of ExperimentActionState with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:status:ExperimentActionState', experimentActionState_status - The state of the action.

$sel:reason:ExperimentActionState', experimentActionState_reason - The reason for the state.

ExperimentState

data ExperimentState

Describes the state of an experiment.

See: newExperimentState smart constructor.

Constructors

ExperimentState' (Maybe ExperimentStatus) (Maybe Text)

newExperimentState :: ExperimentState

Create a value of ExperimentState with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:status:ExperimentState', experimentState_status - The state of the experiment.

$sel:reason:ExperimentState', experimentState_reason - The reason for the state.

ExperimentStopCondition

data ExperimentStopCondition

Describes the stop condition for an experiment.

See: newExperimentStopCondition smart constructor.

Constructors

ExperimentStopCondition' (Maybe Text) (Maybe Text)

newExperimentStopCondition :: ExperimentStopCondition

Create a value of ExperimentStopCondition with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:value:ExperimentStopCondition', experimentStopCondition_value - The Amazon Resource Name (ARN) of the CloudWatch alarm, if applicable.

$sel:source:ExperimentStopCondition', experimentStopCondition_source - The source for the stop condition.

ExperimentSummary

data ExperimentSummary

Provides a summary of an experiment.

See: newExperimentSummary smart constructor.

Constructors

ExperimentSummary' (Maybe POSIX) (Maybe Text) (Maybe ExperimentState) (Maybe Text) (Maybe (HashMap Text Text))

newExperimentSummary :: ExperimentSummary

Create a value of ExperimentSummary with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:creationTime:ExperimentSummary', experimentSummary_creationTime - The time that the experiment was created.

$sel:experimentTemplateId:ExperimentSummary', experimentSummary_experimentTemplateId - The ID of the experiment template.

$sel:state:ExperimentSummary', experimentSummary_state - The state of the experiment.

$sel:id:ExperimentSummary', experimentSummary_id - The ID of the experiment.

$sel:tags:ExperimentSummary', experimentSummary_tags - The tags for the experiment.

ExperimentTarget

data ExperimentTarget

Describes a target for an experiment.

See: newExperimentTarget smart constructor.

Constructors

ExperimentTarget' (Maybe Text) (Maybe (HashMap Text Text)) (Maybe [ExperimentTargetFilter]) (Maybe [Text]) (Maybe Text)

newExperimentTarget :: ExperimentTarget

Create a value of ExperimentTarget with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:resourceType:ExperimentTarget', experimentTarget_resourceType - The resource type.

$sel:resourceTags:ExperimentTarget', experimentTarget_resourceTags - The tags for the target resources.

$sel:filters:ExperimentTarget', experimentTarget_filters - The filters to apply to identify target resources using specific attributes.

$sel:resourceArns:ExperimentTarget', experimentTarget_resourceArns - The Amazon Resource Names (ARNs) of the resources.

$sel:selectionMode:ExperimentTarget', experimentTarget_selectionMode - Scopes the identified resources to a specific count or percentage.

ExperimentTargetFilter

data ExperimentTargetFilter

Describes a filter used for the target resources in an experiment.

See: newExperimentTargetFilter smart constructor.

Constructors

ExperimentTargetFilter' (Maybe Text) (Maybe [Text])

newExperimentTargetFilter :: ExperimentTargetFilter

Create a value of ExperimentTargetFilter with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:path:ExperimentTargetFilter', experimentTargetFilter_path - The attribute path for the filter.

$sel:values:ExperimentTargetFilter', experimentTargetFilter_values - The attribute values for the filter.

ExperimentTemplate

data ExperimentTemplate

Describes an experiment template.

See: newExperimentTemplate smart constructor.

Constructors

ExperimentTemplate' (Maybe POSIX) (Maybe (HashMap Text ExperimentTemplateAction)) (Maybe [ExperimentTemplateStopCondition]) (Maybe (HashMap Text ExperimentTemplateTarget)) (Maybe Text) (Maybe POSIX) (Maybe Text) (Maybe (HashMap Text Text)) (Maybe Text)

newExperimentTemplate :: ExperimentTemplate

Create a value of ExperimentTemplate with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:creationTime:ExperimentTemplate', experimentTemplate_creationTime - The time the experiment template was created.

$sel:actions:ExperimentTemplate', experimentTemplate_actions - The actions for the experiment.

$sel:stopConditions:ExperimentTemplate', experimentTemplate_stopConditions - The stop conditions for the experiment.

$sel:targets:ExperimentTemplate', experimentTemplate_targets - The targets for the experiment.

$sel:id:ExperimentTemplate', experimentTemplate_id - The ID of the experiment template.

$sel:lastUpdateTime:ExperimentTemplate', experimentTemplate_lastUpdateTime - The time the experiment template was last updated.

$sel:description:ExperimentTemplate', experimentTemplate_description - The description for the experiment template.

$sel:tags:ExperimentTemplate', experimentTemplate_tags - The tags for the experiment template.

$sel:roleArn:ExperimentTemplate', experimentTemplate_roleArn - The Amazon Resource Name (ARN) of an IAM role.

ExperimentTemplateAction

data ExperimentTemplateAction

Describes an action for an experiment template.

See: newExperimentTemplateAction smart constructor.

Constructors

ExperimentTemplateAction' (Maybe [Text]) (Maybe Text) (Maybe (HashMap Text Text)) (Maybe (HashMap Text Text)) (Maybe Text)

newExperimentTemplateAction :: ExperimentTemplateAction

Create a value of ExperimentTemplateAction with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:startAfter:ExperimentTemplateAction', experimentTemplateAction_startAfter - The name of the action that must be completed before the current action starts.

$sel:actionId:ExperimentTemplateAction', experimentTemplateAction_actionId - The ID of the action.

$sel:parameters:ExperimentTemplateAction', experimentTemplateAction_parameters - The parameters for the action.

$sel:targets:ExperimentTemplateAction', experimentTemplateAction_targets - The targets for the action.

$sel:description:ExperimentTemplateAction', experimentTemplateAction_description - A description for the action.

ExperimentTemplateStopCondition

data ExperimentTemplateStopCondition

Describes a stop condition for an experiment template.

See: newExperimentTemplateStopCondition smart constructor.

Constructors

ExperimentTemplateStopCondition' (Maybe Text) (Maybe Text)

newExperimentTemplateStopCondition :: ExperimentTemplateStopCondition

Create a value of ExperimentTemplateStopCondition with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:value:ExperimentTemplateStopCondition', experimentTemplateStopCondition_value - The Amazon Resource Name (ARN) of the CloudWatch alarm, if applicable.

$sel:source:ExperimentTemplateStopCondition', experimentTemplateStopCondition_source - The source for the stop condition.

ExperimentTemplateSummary

data ExperimentTemplateSummary

Provides a summary of an experiment template.

See: newExperimentTemplateSummary smart constructor.

Constructors

ExperimentTemplateSummary' (Maybe POSIX) (Maybe Text) (Maybe POSIX) (Maybe Text) (Maybe (HashMap Text Text))

newExperimentTemplateSummary :: ExperimentTemplateSummary

Create a value of ExperimentTemplateSummary with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:creationTime:ExperimentTemplateSummary', experimentTemplateSummary_creationTime - The time that the experiment template was created.

$sel:id:ExperimentTemplateSummary', experimentTemplateSummary_id - The ID of the experiment template.

$sel:lastUpdateTime:ExperimentTemplateSummary', experimentTemplateSummary_lastUpdateTime - The time that the experiment template was last updated.

$sel:description:ExperimentTemplateSummary', experimentTemplateSummary_description - The description of the experiment template.

$sel:tags:ExperimentTemplateSummary', experimentTemplateSummary_tags - The tags for the experiment template.

ExperimentTemplateTarget

data ExperimentTemplateTarget

Describes a target for an experiment template.

See: newExperimentTemplateTarget smart constructor.

Constructors

ExperimentTemplateTarget' (Maybe Text) (Maybe (HashMap Text Text)) (Maybe [ExperimentTemplateTargetFilter]) (Maybe [Text]) (Maybe Text)

newExperimentTemplateTarget :: ExperimentTemplateTarget

Create a value of ExperimentTemplateTarget with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:resourceType:ExperimentTemplateTarget', experimentTemplateTarget_resourceType - The resource type.

$sel:resourceTags:ExperimentTemplateTarget', experimentTemplateTarget_resourceTags - The tags for the target resources.

$sel:filters:ExperimentTemplateTarget', experimentTemplateTarget_filters - The filters to apply to identify target resources using specific attributes.

$sel:resourceArns:ExperimentTemplateTarget', experimentTemplateTarget_resourceArns - The Amazon Resource Names (ARNs) of the targets.

$sel:selectionMode:ExperimentTemplateTarget', experimentTemplateTarget_selectionMode - Scopes the identified resources to a specific count or percentage.

ExperimentTemplateTargetFilter

data ExperimentTemplateTargetFilter

Describes a filter used for the target resources in an experiment template.

See: newExperimentTemplateTargetFilter smart constructor.

Constructors

ExperimentTemplateTargetFilter' (Maybe Text) (Maybe [Text])

newExperimentTemplateTargetFilter :: ExperimentTemplateTargetFilter

Create a value of ExperimentTemplateTargetFilter with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:path:ExperimentTemplateTargetFilter', experimentTemplateTargetFilter_path - The attribute path for the filter.

$sel:values:ExperimentTemplateTargetFilter', experimentTemplateTargetFilter_values - The attribute values for the filter.

ExperimentTemplateTargetInputFilter

data ExperimentTemplateTargetInputFilter

Describes a filter used for the target resource input in an experiment template.

See: newExperimentTemplateTargetInputFilter smart constructor.

Constructors

ExperimentTemplateTargetInputFilter' Text [Text]

newExperimentTemplateTargetInputFilter

Arguments

:: Text
$sel:path:ExperimentTemplateTargetInputFilter'

-> ExperimentTemplateTargetInputFilter

Create a value of ExperimentTemplateTargetInputFilter with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:path:ExperimentTemplateTargetInputFilter', experimentTemplateTargetInputFilter_path - The attribute path for the filter.

$sel:values:ExperimentTemplateTargetInputFilter', experimentTemplateTargetInputFilter_values - The attribute values for the filter.

UpdateExperimentTemplateActionInputItem

data UpdateExperimentTemplateActionInputItem

Specifies an action for an experiment template.

See: newUpdateExperimentTemplateActionInputItem smart constructor.

Constructors

UpdateExperimentTemplateActionInputItem' (Maybe [Text]) (Maybe Text) (Maybe (HashMap Text Text)) (Maybe (HashMap Text Text)) (Maybe Text)

newUpdateExperimentTemplateActionInputItem :: UpdateExperimentTemplateActionInputItem

Create a value of UpdateExperimentTemplateActionInputItem with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:startAfter:UpdateExperimentTemplateActionInputItem', updateExperimentTemplateActionInputItem_startAfter - The name of the action that must be completed before the current action starts. Omit this parameter to run the action at the start of the experiment.

$sel:actionId:UpdateExperimentTemplateActionInputItem', updateExperimentTemplateActionInputItem_actionId - The ID of the action.

$sel:parameters:UpdateExperimentTemplateActionInputItem', updateExperimentTemplateActionInputItem_parameters - The parameters for the action, if applicable.

$sel:targets:UpdateExperimentTemplateActionInputItem', updateExperimentTemplateActionInputItem_targets - The targets for the action.

$sel:description:UpdateExperimentTemplateActionInputItem', updateExperimentTemplateActionInputItem_description - A description for the action.

UpdateExperimentTemplateStopConditionInput

data UpdateExperimentTemplateStopConditionInput

Specifies a stop condition for an experiment. You can define a stop condition as a CloudWatch alarm.

See: newUpdateExperimentTemplateStopConditionInput smart constructor.

Constructors

UpdateExperimentTemplateStopConditionInput' (Maybe Text) Text

newUpdateExperimentTemplateStopConditionInput

Arguments

:: Text
$sel:source:UpdateExperimentTemplateStopConditionInput'

-> UpdateExperimentTemplateStopConditionInput

Create a value of UpdateExperimentTemplateStopConditionInput with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:value:UpdateExperimentTemplateStopConditionInput', updateExperimentTemplateStopConditionInput_value - The Amazon Resource Name (ARN) of the CloudWatch alarm.

$sel:source:UpdateExperimentTemplateStopConditionInput', updateExperimentTemplateStopConditionInput_source - The source for the stop condition. Specify aws:cloudwatch:alarm if the stop condition is defined by a CloudWatch alarm. Specify none if there is no stop condition.

UpdateExperimentTemplateTargetInput

data UpdateExperimentTemplateTargetInput

Specifies a target for an experiment. You must specify at least one Amazon Resource Name (ARN) or at least one resource tag. You cannot specify both.

See: newUpdateExperimentTemplateTargetInput smart constructor.

Constructors

UpdateExperimentTemplateTargetInput' (Maybe (HashMap Text Text)) (Maybe [ExperimentTemplateTargetInputFilter]) (Maybe [Text]) Text Text

newUpdateExperimentTemplateTargetInput

Arguments

:: Text
$sel:resourceType:UpdateExperimentTemplateTargetInput'

-> Text
$sel:selectionMode:UpdateExperimentTemplateTargetInput'

-> UpdateExperimentTemplateTargetInput

Create a value of UpdateExperimentTemplateTargetInput with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:resourceTags:UpdateExperimentTemplateTargetInput', updateExperimentTemplateTargetInput_resourceTags - The tags for the target resources.

$sel:filters:UpdateExperimentTemplateTargetInput', updateExperimentTemplateTargetInput_filters - The filters to apply to identify target resources using specific attributes.

$sel:resourceArns:UpdateExperimentTemplateTargetInput', updateExperimentTemplateTargetInput_resourceArns - The Amazon Resource Names (ARNs) of the targets.

$sel:resourceType:UpdateExperimentTemplateTargetInput', updateExperimentTemplateTargetInput_resourceType - The AWS resource type. The resource type must be supported for the specified action.

$sel:selectionMode:UpdateExperimentTemplateTargetInput', updateExperimentTemplateTargetInput_selectionMode - Scopes the identified resources to a specific count or percentage.