Amazonka.FIS.CreateExperimentTemplate

Description

Creates an experiment template.

To create a template, specify the following information:

Targets: A target can be a specific resource in your AWS environment, or one or more resources that match criteria that you specify, for example, resources that have specific tags.

Actions: The actions to carry out on the target. You can specify multiple actions, the duration of each action, and when to start each action during an experiment.

Stop conditions: If a stop condition is triggered while an experiment is running, the experiment is automatically stopped. You can define a stop condition as a CloudWatch alarm.

For more information, see the AWS Fault Injection Simulator User Guide.

Creating a Request

data CreateExperimentTemplate

See: newCreateExperimentTemplate smart constructor.

Constructors

CreateExperimentTemplate'

Fields

targets :: Maybe (HashMap Text CreateExperimentTemplateTargetInput)

The targets for the experiment.

tags :: Maybe (HashMap Text Text)

The tags to apply to the experiment template.

clientToken :: Text

Unique, case-sensitive identifier that you provide to ensure the idempotency of the request.

description :: Text

A description for the experiment template. Can contain up to 64 letters (A-Z and a-z).

stopConditions :: [CreateExperimentTemplateStopConditionInput]

The stop conditions.

actions :: HashMap Text CreateExperimentTemplateActionInput

The actions for the experiment.

roleArn :: Text

The Amazon Resource Name (ARN) of an IAM role that grants the AWS FIS service permission to perform service actions on your behalf.

newCreateExperimentTemplate

Arguments

:: Text
$sel:clientToken:CreateExperimentTemplate'

-> Text
$sel:description:CreateExperimentTemplate'

-> Text
$sel:roleArn:CreateExperimentTemplate'

-> CreateExperimentTemplate

Create a value of CreateExperimentTemplate with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:targets:CreateExperimentTemplate', createExperimentTemplate_targets - The targets for the experiment.

$sel:tags:CreateExperimentTemplate', createExperimentTemplate_tags - The tags to apply to the experiment template.

$sel:clientToken:CreateExperimentTemplate', createExperimentTemplate_clientToken - Unique, case-sensitive identifier that you provide to ensure the idempotency of the request.

$sel:description:CreateExperimentTemplate', createExperimentTemplate_description - A description for the experiment template. Can contain up to 64 letters (A-Z and a-z).

$sel:stopConditions:CreateExperimentTemplate', createExperimentTemplate_stopConditions - The stop conditions.

$sel:actions:CreateExperimentTemplate', createExperimentTemplate_actions - The actions for the experiment.

$sel:roleArn:CreateExperimentTemplate', createExperimentTemplate_roleArn - The Amazon Resource Name (ARN) of an IAM role that grants the AWS FIS service permission to perform service actions on your behalf.

Request Lenses

createExperimentTemplate_targets :: Lens' CreateExperimentTemplate (Maybe (HashMap Text CreateExperimentTemplateTargetInput))

The targets for the experiment.

createExperimentTemplate_tags :: Lens' CreateExperimentTemplate (Maybe (HashMap Text Text))

The tags to apply to the experiment template.

createExperimentTemplate_clientToken :: Lens' CreateExperimentTemplate Text

Unique, case-sensitive identifier that you provide to ensure the idempotency of the request.

createExperimentTemplate_description :: Lens' CreateExperimentTemplate Text

A description for the experiment template. Can contain up to 64 letters (A-Z and a-z).

createExperimentTemplate_stopConditions :: Lens' CreateExperimentTemplate [CreateExperimentTemplateStopConditionInput]

The stop conditions.

createExperimentTemplate_actions :: Lens' CreateExperimentTemplate (HashMap Text CreateExperimentTemplateActionInput)

The actions for the experiment.

createExperimentTemplate_roleArn :: Lens' CreateExperimentTemplate Text

The Amazon Resource Name (ARN) of an IAM role that grants the AWS FIS service permission to perform service actions on your behalf.

Destructuring the Response

data CreateExperimentTemplateResponse

See: newCreateExperimentTemplateResponse smart constructor.

Constructors

CreateExperimentTemplateResponse'

Fields

experimentTemplate :: Maybe ExperimentTemplate

Information about the experiment template.

httpStatus :: Int

The response's http status code.

newCreateExperimentTemplateResponse

Arguments

:: Int
$sel:httpStatus:CreateExperimentTemplateResponse'

-> CreateExperimentTemplateResponse

Create a value of CreateExperimentTemplateResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:experimentTemplate:CreateExperimentTemplateResponse', createExperimentTemplateResponse_experimentTemplate - Information about the experiment template.

$sel:httpStatus:CreateExperimentTemplateResponse', createExperimentTemplateResponse_httpStatus - The response's http status code.

Response Lenses

createExperimentTemplateResponse_experimentTemplate :: Lens' CreateExperimentTemplateResponse (Maybe ExperimentTemplate)

Information about the experiment template.

createExperimentTemplateResponse_httpStatus :: Lens' CreateExperimentTemplateResponse Int

The response's http status code.