Amazonka.FIS.Lens

Description

Operations

GetExperimentTemplate

getExperimentTemplate_id :: Lens' GetExperimentTemplate Text

The ID of the experiment template.

getExperimentTemplateResponse_experimentTemplate :: Lens' GetExperimentTemplateResponse (Maybe ExperimentTemplate)

Information about the experiment template.

getExperimentTemplateResponse_httpStatus :: Lens' GetExperimentTemplateResponse Int

The response's http status code.

ListActions

listActions_nextToken :: Lens' ListActions (Maybe Text)

The token for the next page of results.

listActions_maxResults :: Lens' ListActions (Maybe Natural)

The maximum number of results to return with a single call. To retrieve the remaining results, make another call with the returned nextToken value.

listActionsResponse_actions :: Lens' ListActionsResponse (Maybe [ActionSummary])

The actions.

listActionsResponse_nextToken :: Lens' ListActionsResponse (Maybe Text)

The token to use to retrieve the next page of results. This value is null when there are no more results to return.

listActionsResponse_httpStatus :: Lens' ListActionsResponse Int

The response's http status code.

ListTagsForResource

listTagsForResource_resourceArn :: Lens' ListTagsForResource Text

The Amazon Resource Name (ARN) of the resource.

listTagsForResourceResponse_tags :: Lens' ListTagsForResourceResponse (Maybe (HashMap Text Text))

The tags for the resource.

listTagsForResourceResponse_httpStatus :: Lens' ListTagsForResourceResponse Int

The response's http status code.

CreateExperimentTemplate

createExperimentTemplate_targets :: Lens' CreateExperimentTemplate (Maybe (HashMap Text CreateExperimentTemplateTargetInput))

The targets for the experiment.

createExperimentTemplate_tags :: Lens' CreateExperimentTemplate (Maybe (HashMap Text Text))

The tags to apply to the experiment template.

createExperimentTemplate_clientToken :: Lens' CreateExperimentTemplate Text

Unique, case-sensitive identifier that you provide to ensure the idempotency of the request.

createExperimentTemplate_description :: Lens' CreateExperimentTemplate Text

A description for the experiment template. Can contain up to 64 letters (A-Z and a-z).

createExperimentTemplate_stopConditions :: Lens' CreateExperimentTemplate [CreateExperimentTemplateStopConditionInput]

The stop conditions.

createExperimentTemplate_actions :: Lens' CreateExperimentTemplate (HashMap Text CreateExperimentTemplateActionInput)

The actions for the experiment.

createExperimentTemplate_roleArn :: Lens' CreateExperimentTemplate Text

The Amazon Resource Name (ARN) of an IAM role that grants the AWS FIS service permission to perform service actions on your behalf.

createExperimentTemplateResponse_experimentTemplate :: Lens' CreateExperimentTemplateResponse (Maybe ExperimentTemplate)

Information about the experiment template.

createExperimentTemplateResponse_httpStatus :: Lens' CreateExperimentTemplateResponse Int

The response's http status code.

ListExperiments

listExperiments_nextToken :: Lens' ListExperiments (Maybe Text)

The token for the next page of results.

listExperiments_maxResults :: Lens' ListExperiments (Maybe Natural)

The maximum number of results to return with a single call. To retrieve the remaining results, make another call with the returned nextToken value.

listExperimentsResponse_experiments :: Lens' ListExperimentsResponse (Maybe [ExperimentSummary])

The experiments.

listExperimentsResponse_nextToken :: Lens' ListExperimentsResponse (Maybe Text)

The token to use to retrieve the next page of results. This value is null when there are no more results to return.

listExperimentsResponse_httpStatus :: Lens' ListExperimentsResponse Int

The response's http status code.

UpdateExperimentTemplate

updateExperimentTemplate_actions :: Lens' UpdateExperimentTemplate (Maybe (HashMap Text UpdateExperimentTemplateActionInputItem))

The actions for the experiment.

updateExperimentTemplate_stopConditions :: Lens' UpdateExperimentTemplate (Maybe [UpdateExperimentTemplateStopConditionInput])

The stop conditions for the experiment.

updateExperimentTemplate_targets :: Lens' UpdateExperimentTemplate (Maybe (HashMap Text UpdateExperimentTemplateTargetInput))

The targets for the experiment.

updateExperimentTemplate_description :: Lens' UpdateExperimentTemplate (Maybe Text)

A description for the template.

updateExperimentTemplate_roleArn :: Lens' UpdateExperimentTemplate (Maybe Text)

The Amazon Resource Name (ARN) of an IAM role that grants the AWS FIS service permission to perform service actions on your behalf.

updateExperimentTemplate_id :: Lens' UpdateExperimentTemplate Text

The ID of the experiment template.

updateExperimentTemplateResponse_experimentTemplate :: Lens' UpdateExperimentTemplateResponse (Maybe ExperimentTemplate)

Information about the experiment template.

updateExperimentTemplateResponse_httpStatus :: Lens' UpdateExperimentTemplateResponse Int

The response's http status code.

GetAction

getAction_id :: Lens' GetAction Text

The ID of the action.

getActionResponse_action :: Lens' GetActionResponse (Maybe Action)

Information about the action.

getActionResponse_httpStatus :: Lens' GetActionResponse Int

The response's http status code.

DeleteExperimentTemplate

deleteExperimentTemplate_id :: Lens' DeleteExperimentTemplate Text

The ID of the experiment template.

deleteExperimentTemplateResponse_experimentTemplate :: Lens' DeleteExperimentTemplateResponse (Maybe ExperimentTemplate)

Information about the experiment template.

deleteExperimentTemplateResponse_httpStatus :: Lens' DeleteExperimentTemplateResponse Int

The response's http status code.

StartExperiment

startExperiment_tags :: Lens' StartExperiment (Maybe (HashMap Text Text))

The tags to apply to the experiment.

startExperiment_clientToken :: Lens' StartExperiment Text

Unique, case-sensitive identifier that you provide to ensure the idempotency of the request.

startExperiment_experimentTemplateId :: Lens' StartExperiment Text

The ID of the experiment template.

startExperimentResponse_experiment :: Lens' StartExperimentResponse (Maybe Experiment)

Information about the experiment.

startExperimentResponse_httpStatus :: Lens' StartExperimentResponse Int

The response's http status code.

GetExperiment

getExperiment_id :: Lens' GetExperiment Text

The ID of the experiment.

getExperimentResponse_experiment :: Lens' GetExperimentResponse (Maybe Experiment)

Information about the experiment.

getExperimentResponse_httpStatus :: Lens' GetExperimentResponse Int

The response's http status code.

ListExperimentTemplates

listExperimentTemplates_nextToken :: Lens' ListExperimentTemplates (Maybe Text)

The token for the next page of results.

listExperimentTemplates_maxResults :: Lens' ListExperimentTemplates (Maybe Natural)

The maximum number of results to return with a single call. To retrieve the remaining results, make another call with the returned nextToken value.

listExperimentTemplatesResponse_nextToken :: Lens' ListExperimentTemplatesResponse (Maybe Text)

The token to use to retrieve the next page of results. This value is null when there are no more results to return.

listExperimentTemplatesResponse_experimentTemplates :: Lens' ListExperimentTemplatesResponse (Maybe [ExperimentTemplateSummary])

The experiment templates.

listExperimentTemplatesResponse_httpStatus :: Lens' ListExperimentTemplatesResponse Int

The response's http status code.

TagResource

tagResource_resourceArn :: Lens' TagResource Text

The Amazon Resource Name (ARN) of the resource.

tagResource_tags :: Lens' TagResource (HashMap Text Text)

The tags for the resource.

tagResourceResponse_httpStatus :: Lens' TagResourceResponse Int

The response's http status code.

UntagResource

untagResource_tagKeys :: Lens' UntagResource (Maybe [Text])

The tag keys to remove.

untagResource_resourceArn :: Lens' UntagResource Text

The Amazon Resource Name (ARN) of the resource.

untagResourceResponse_httpStatus :: Lens' UntagResourceResponse Int

The response's http status code.

StopExperiment

stopExperiment_id :: Lens' StopExperiment Text

The ID of the experiment.

stopExperimentResponse_experiment :: Lens' StopExperimentResponse (Maybe Experiment)

Information about the experiment.

stopExperimentResponse_httpStatus :: Lens' StopExperimentResponse Int

The response's http status code.

Types

Action

action_parameters :: Lens' Action (Maybe (HashMap Text ActionParameter))

The action parameters, if applicable.

action_targets :: Lens' Action (Maybe (HashMap Text ActionTarget))

The supported targets for the action.

action_id :: Lens' Action (Maybe Text)

The ID of the action.

action_description :: Lens' Action (Maybe Text)

The description for the action.

action_tags :: Lens' Action (Maybe (HashMap Text Text))

The tags for the action.

ActionParameter

actionParameter_required :: Lens' ActionParameter (Maybe Bool)

Indicates whether the parameter is required.

actionParameter_description :: Lens' ActionParameter (Maybe Text)

The parameter description.

ActionSummary

actionSummary_targets :: Lens' ActionSummary (Maybe (HashMap Text ActionTarget))

The targets for the action.

actionSummary_id :: Lens' ActionSummary (Maybe Text)

The ID of the action.

actionSummary_description :: Lens' ActionSummary (Maybe Text)

The description for the action.

actionSummary_tags :: Lens' ActionSummary (Maybe (HashMap Text Text))

The tags for the action.

ActionTarget

actionTarget_resourceType :: Lens' ActionTarget (Maybe Text)

The resource type of the target.

CreateExperimentTemplateActionInput

createExperimentTemplateActionInput_startAfter :: Lens' CreateExperimentTemplateActionInput (Maybe [Text])

The name of the action that must be completed before the current action starts. Omit this parameter to run the action at the start of the experiment.

createExperimentTemplateActionInput_parameters :: Lens' CreateExperimentTemplateActionInput (Maybe (HashMap Text Text))

The parameters for the action, if applicable.

createExperimentTemplateActionInput_targets :: Lens' CreateExperimentTemplateActionInput (Maybe (HashMap Text Text))

The targets for the action.

createExperimentTemplateActionInput_description :: Lens' CreateExperimentTemplateActionInput (Maybe Text)

A description for the action.

createExperimentTemplateActionInput_actionId :: Lens' CreateExperimentTemplateActionInput Text

The ID of the action.

CreateExperimentTemplateStopConditionInput

createExperimentTemplateStopConditionInput_value :: Lens' CreateExperimentTemplateStopConditionInput (Maybe Text)

The Amazon Resource Name (ARN) of the CloudWatch alarm. This is required if the source is a CloudWatch alarm.

createExperimentTemplateStopConditionInput_source :: Lens' CreateExperimentTemplateStopConditionInput Text

The source for the stop condition. Specify aws:cloudwatch:alarm if the stop condition is defined by a CloudWatch alarm. Specify none if there is no stop condition.

CreateExperimentTemplateTargetInput

createExperimentTemplateTargetInput_resourceTags :: Lens' CreateExperimentTemplateTargetInput (Maybe (HashMap Text Text))

The tags for the target resources.

createExperimentTemplateTargetInput_filters :: Lens' CreateExperimentTemplateTargetInput (Maybe [ExperimentTemplateTargetInputFilter])

The filters to apply to identify target resources using specific attributes.

createExperimentTemplateTargetInput_resourceArns :: Lens' CreateExperimentTemplateTargetInput (Maybe [Text])

The Amazon Resource Names (ARNs) of the resources.

createExperimentTemplateTargetInput_resourceType :: Lens' CreateExperimentTemplateTargetInput Text

The AWS resource type. The resource type must be supported for the specified action.

createExperimentTemplateTargetInput_selectionMode :: Lens' CreateExperimentTemplateTargetInput Text

Scopes the identified resources to a specific count of the resources at random, or a percentage of the resources. All identified resources are included in the target.

ALL - Run the action on all identified targets. This is the default.

COUNT(n) - Run the action on the specified number of targets, chosen from the identified targets at random. For example, COUNT(1) selects one of the targets.

PERCENT(n) - Run the action on the specified percentage of targets, chosen from the identified targets at random. For example, PERCENT(25) selects 25% of the targets.

Experiment

experiment_creationTime :: Lens' Experiment (Maybe UTCTime)

The time the experiment was created.

experiment_experimentTemplateId :: Lens' Experiment (Maybe Text)

The ID of the experiment template.

experiment_state :: Lens' Experiment (Maybe ExperimentState)

The state of the experiment.

experiment_actions :: Lens' Experiment (Maybe (HashMap Text ExperimentAction))

The actions for the experiment.

experiment_startTime :: Lens' Experiment (Maybe UTCTime)

The time that the experiment was started.

experiment_stopConditions :: Lens' Experiment (Maybe [ExperimentStopCondition])

The stop conditions for the experiment.

experiment_endTime :: Lens' Experiment (Maybe UTCTime)

The time that the experiment ended.

experiment_targets :: Lens' Experiment (Maybe (HashMap Text ExperimentTarget))

The targets for the experiment.

experiment_id :: Lens' Experiment (Maybe Text)

The ID of the experiment.

experiment_tags :: Lens' Experiment (Maybe (HashMap Text Text))

The tags for the experiment.

experiment_roleArn :: Lens' Experiment (Maybe Text)

The Amazon Resource Name (ARN) of an IAM role that grants the AWS FIS service permission to perform service actions on your behalf.

ExperimentAction

experimentAction_startAfter :: Lens' ExperimentAction (Maybe [Text])

The name of the action that must be completed before this action starts.

experimentAction_state :: Lens' ExperimentAction (Maybe ExperimentActionState)

The state of the action.

experimentAction_actionId :: Lens' ExperimentAction (Maybe Text)

The ID of the action.

experimentAction_parameters :: Lens' ExperimentAction (Maybe (HashMap Text Text))

The parameters for the action.

experimentAction_targets :: Lens' ExperimentAction (Maybe (HashMap Text Text))

The targets for the action.

experimentAction_description :: Lens' ExperimentAction (Maybe Text)

The description for the action.

ExperimentActionState

experimentActionState_status :: Lens' ExperimentActionState (Maybe ExperimentActionStatus)

The state of the action.

experimentActionState_reason :: Lens' ExperimentActionState (Maybe Text)

The reason for the state.

ExperimentState

experimentState_status :: Lens' ExperimentState (Maybe ExperimentStatus)

The state of the experiment.

experimentState_reason :: Lens' ExperimentState (Maybe Text)

The reason for the state.

ExperimentStopCondition

experimentStopCondition_value :: Lens' ExperimentStopCondition (Maybe Text)

The Amazon Resource Name (ARN) of the CloudWatch alarm, if applicable.

experimentStopCondition_source :: Lens' ExperimentStopCondition (Maybe Text)

The source for the stop condition.

ExperimentSummary

experimentSummary_creationTime :: Lens' ExperimentSummary (Maybe UTCTime)

The time that the experiment was created.

experimentSummary_experimentTemplateId :: Lens' ExperimentSummary (Maybe Text)

The ID of the experiment template.

experimentSummary_state :: Lens' ExperimentSummary (Maybe ExperimentState)

The state of the experiment.

experimentSummary_id :: Lens' ExperimentSummary (Maybe Text)

The ID of the experiment.

experimentSummary_tags :: Lens' ExperimentSummary (Maybe (HashMap Text Text))

The tags for the experiment.

ExperimentTarget

experimentTarget_resourceType :: Lens' ExperimentTarget (Maybe Text)

The resource type.

experimentTarget_resourceTags :: Lens' ExperimentTarget (Maybe (HashMap Text Text))

The tags for the target resources.

experimentTarget_filters :: Lens' ExperimentTarget (Maybe [ExperimentTargetFilter])

The filters to apply to identify target resources using specific attributes.

experimentTarget_resourceArns :: Lens' ExperimentTarget (Maybe [Text])

The Amazon Resource Names (ARNs) of the resources.

experimentTarget_selectionMode :: Lens' ExperimentTarget (Maybe Text)

Scopes the identified resources to a specific count or percentage.

ExperimentTargetFilter

experimentTargetFilter_path :: Lens' ExperimentTargetFilter (Maybe Text)

The attribute path for the filter.

experimentTargetFilter_values :: Lens' ExperimentTargetFilter (Maybe [Text])

The attribute values for the filter.

ExperimentTemplate

experimentTemplate_creationTime :: Lens' ExperimentTemplate (Maybe UTCTime)

The time the experiment template was created.

experimentTemplate_actions :: Lens' ExperimentTemplate (Maybe (HashMap Text ExperimentTemplateAction))

The actions for the experiment.

experimentTemplate_stopConditions :: Lens' ExperimentTemplate (Maybe [ExperimentTemplateStopCondition])

The stop conditions for the experiment.

experimentTemplate_targets :: Lens' ExperimentTemplate (Maybe (HashMap Text ExperimentTemplateTarget))

The targets for the experiment.

experimentTemplate_id :: Lens' ExperimentTemplate (Maybe Text)

The ID of the experiment template.

experimentTemplate_lastUpdateTime :: Lens' ExperimentTemplate (Maybe UTCTime)

The time the experiment template was last updated.

experimentTemplate_description :: Lens' ExperimentTemplate (Maybe Text)

The description for the experiment template.

experimentTemplate_tags :: Lens' ExperimentTemplate (Maybe (HashMap Text Text))

The tags for the experiment template.

experimentTemplate_roleArn :: Lens' ExperimentTemplate (Maybe Text)

The Amazon Resource Name (ARN) of an IAM role.

ExperimentTemplateAction

experimentTemplateAction_startAfter :: Lens' ExperimentTemplateAction (Maybe [Text])

The name of the action that must be completed before the current action starts.

experimentTemplateAction_actionId :: Lens' ExperimentTemplateAction (Maybe Text)

The ID of the action.

experimentTemplateAction_parameters :: Lens' ExperimentTemplateAction (Maybe (HashMap Text Text))

The parameters for the action.

experimentTemplateAction_targets :: Lens' ExperimentTemplateAction (Maybe (HashMap Text Text))

The targets for the action.

experimentTemplateAction_description :: Lens' ExperimentTemplateAction (Maybe Text)

A description for the action.

ExperimentTemplateStopCondition

experimentTemplateStopCondition_value :: Lens' ExperimentTemplateStopCondition (Maybe Text)

The Amazon Resource Name (ARN) of the CloudWatch alarm, if applicable.

experimentTemplateStopCondition_source :: Lens' ExperimentTemplateStopCondition (Maybe Text)

The source for the stop condition.

ExperimentTemplateSummary

experimentTemplateSummary_creationTime :: Lens' ExperimentTemplateSummary (Maybe UTCTime)

The time that the experiment template was created.

experimentTemplateSummary_id :: Lens' ExperimentTemplateSummary (Maybe Text)

The ID of the experiment template.

experimentTemplateSummary_lastUpdateTime :: Lens' ExperimentTemplateSummary (Maybe UTCTime)

The time that the experiment template was last updated.

experimentTemplateSummary_description :: Lens' ExperimentTemplateSummary (Maybe Text)

The description of the experiment template.

experimentTemplateSummary_tags :: Lens' ExperimentTemplateSummary (Maybe (HashMap Text Text))

The tags for the experiment template.

ExperimentTemplateTarget

experimentTemplateTarget_resourceType :: Lens' ExperimentTemplateTarget (Maybe Text)

The resource type.

experimentTemplateTarget_resourceTags :: Lens' ExperimentTemplateTarget (Maybe (HashMap Text Text))

The tags for the target resources.

experimentTemplateTarget_filters :: Lens' ExperimentTemplateTarget (Maybe [ExperimentTemplateTargetFilter])

The filters to apply to identify target resources using specific attributes.

experimentTemplateTarget_resourceArns :: Lens' ExperimentTemplateTarget (Maybe [Text])

The Amazon Resource Names (ARNs) of the targets.

experimentTemplateTarget_selectionMode :: Lens' ExperimentTemplateTarget (Maybe Text)

Scopes the identified resources to a specific count or percentage.

ExperimentTemplateTargetFilter

experimentTemplateTargetFilter_path :: Lens' ExperimentTemplateTargetFilter (Maybe Text)

The attribute path for the filter.

experimentTemplateTargetFilter_values :: Lens' ExperimentTemplateTargetFilter (Maybe [Text])

The attribute values for the filter.

ExperimentTemplateTargetInputFilter

experimentTemplateTargetInputFilter_path :: Lens' ExperimentTemplateTargetInputFilter Text

The attribute path for the filter.

experimentTemplateTargetInputFilter_values :: Lens' ExperimentTemplateTargetInputFilter [Text]

The attribute values for the filter.

UpdateExperimentTemplateActionInputItem

updateExperimentTemplateActionInputItem_startAfter :: Lens' UpdateExperimentTemplateActionInputItem (Maybe [Text])

The name of the action that must be completed before the current action starts. Omit this parameter to run the action at the start of the experiment.

updateExperimentTemplateActionInputItem_actionId :: Lens' UpdateExperimentTemplateActionInputItem (Maybe Text)

The ID of the action.

updateExperimentTemplateActionInputItem_parameters :: Lens' UpdateExperimentTemplateActionInputItem (Maybe (HashMap Text Text))

The parameters for the action, if applicable.

updateExperimentTemplateActionInputItem_targets :: Lens' UpdateExperimentTemplateActionInputItem (Maybe (HashMap Text Text))

The targets for the action.

updateExperimentTemplateActionInputItem_description :: Lens' UpdateExperimentTemplateActionInputItem (Maybe Text)

A description for the action.

UpdateExperimentTemplateStopConditionInput

updateExperimentTemplateStopConditionInput_value :: Lens' UpdateExperimentTemplateStopConditionInput (Maybe Text)

The Amazon Resource Name (ARN) of the CloudWatch alarm.

updateExperimentTemplateStopConditionInput_source :: Lens' UpdateExperimentTemplateStopConditionInput Text

The source for the stop condition. Specify aws:cloudwatch:alarm if the stop condition is defined by a CloudWatch alarm. Specify none if there is no stop condition.

UpdateExperimentTemplateTargetInput

updateExperimentTemplateTargetInput_resourceTags :: Lens' UpdateExperimentTemplateTargetInput (Maybe (HashMap Text Text))

The tags for the target resources.

updateExperimentTemplateTargetInput_filters :: Lens' UpdateExperimentTemplateTargetInput (Maybe [ExperimentTemplateTargetInputFilter])

The filters to apply to identify target resources using specific attributes.

updateExperimentTemplateTargetInput_resourceArns :: Lens' UpdateExperimentTemplateTargetInput (Maybe [Text])

The Amazon Resource Names (ARNs) of the targets.

updateExperimentTemplateTargetInput_resourceType :: Lens' UpdateExperimentTemplateTargetInput Text

The AWS resource type. The resource type must be supported for the specified action.

updateExperimentTemplateTargetInput_selectionMode :: Lens' UpdateExperimentTemplateTargetInput Text

Scopes the identified resources to a specific count or percentage.