Amazonka.FIS.ListActions

Description

Lists the available AWS FIS actions.

Creating a Request

data ListActions

See: newListActions smart constructor.

Constructors

ListActions'

Fields

nextToken :: Maybe Text

The token for the next page of results.

maxResults :: Maybe Natural

The maximum number of results to return with a single call. To retrieve the remaining results, make another call with the returned nextToken value.

newListActions :: ListActions

Create a value of ListActions with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:nextToken:ListActions', listActions_nextToken - The token for the next page of results.

$sel:maxResults:ListActions', listActions_maxResults - The maximum number of results to return with a single call. To retrieve the remaining results, make another call with the returned nextToken value.

Request Lenses

listActions_nextToken :: Lens' ListActions (Maybe Text)

The token for the next page of results.

listActions_maxResults :: Lens' ListActions (Maybe Natural)

The maximum number of results to return with a single call. To retrieve the remaining results, make another call with the returned nextToken value.

Destructuring the Response

data ListActionsResponse

See: newListActionsResponse smart constructor.

Constructors

ListActionsResponse'

Fields

actions :: Maybe [ActionSummary]

The actions.

nextToken :: Maybe Text

The token to use to retrieve the next page of results. This value is null when there are no more results to return.

httpStatus :: Int

The response's http status code.

newListActionsResponse

Arguments

:: Int
$sel:httpStatus:ListActionsResponse'

-> ListActionsResponse

Create a value of ListActionsResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:actions:ListActionsResponse', listActionsResponse_actions - The actions.

$sel:nextToken:ListActions', listActionsResponse_nextToken - The token to use to retrieve the next page of results. This value is null when there are no more results to return.

$sel:httpStatus:ListActionsResponse', listActionsResponse_httpStatus - The response's http status code.

Response Lenses

listActionsResponse_actions :: Lens' ListActionsResponse (Maybe [ActionSummary])

The actions.

listActionsResponse_nextToken :: Lens' ListActionsResponse (Maybe Text)

The token to use to retrieve the next page of results. This value is null when there are no more results to return.

listActionsResponse_httpStatus :: Lens' ListActionsResponse Int

The response's http status code.