Amazonka.FIS.Types

Description

Service Configuration

defaultService :: Service

API version 2020-12-01 of the Amazon Fault Injection Simulator SDK configuration.

Errors

_ValidationException :: AsError a => Getting (First ServiceError) a ServiceError

The specified input is not valid, or fails to satisfy the constraints for the request.

_ConflictException :: AsError a => Getting (First ServiceError) a ServiceError

The request could not be processed because of a conflict.

_ServiceQuotaExceededException :: AsError a => Getting (First ServiceError) a ServiceError

You have exceeded your service quota.

_ResourceNotFoundException :: AsError a => Getting (First ServiceError) a ServiceError

The specified resource cannot be found.

ExperimentActionStatus

newtype ExperimentActionStatus

Constructors

ExperimentActionStatus'

Fields

fromExperimentActionStatus :: Text

Bundled Patterns

pattern ExperimentActionStatus_Cancelled :: ExperimentActionStatus

pattern ExperimentActionStatus_Completed :: ExperimentActionStatus

pattern ExperimentActionStatus_Failed :: ExperimentActionStatus

pattern ExperimentActionStatus_Initiating :: ExperimentActionStatus

pattern ExperimentActionStatus_Pending :: ExperimentActionStatus

pattern ExperimentActionStatus_Running :: ExperimentActionStatus

pattern ExperimentActionStatus_Stopped :: ExperimentActionStatus

pattern ExperimentActionStatus_Stopping :: ExperimentActionStatus

ExperimentStatus

newtype ExperimentStatus

Constructors

ExperimentStatus'

Fields

fromExperimentStatus :: Text

Bundled Patterns

pattern ExperimentStatus_Completed :: ExperimentStatus

pattern ExperimentStatus_Failed :: ExperimentStatus

pattern ExperimentStatus_Initiating :: ExperimentStatus

pattern ExperimentStatus_Pending :: ExperimentStatus

pattern ExperimentStatus_Running :: ExperimentStatus

pattern ExperimentStatus_Stopped :: ExperimentStatus

pattern ExperimentStatus_Stopping :: ExperimentStatus

Action

data Action

Describes an action. For more information, see AWS FIS actions in the AWS Fault Injection Simulator User Guide.

See: newAction smart constructor.

Constructors

Action'

Fields

parameters :: Maybe (HashMap Text ActionParameter)

The action parameters, if applicable.

targets :: Maybe (HashMap Text ActionTarget)

The supported targets for the action.

id :: Maybe Text

The ID of the action.

description :: Maybe Text

The description for the action.

tags :: Maybe (HashMap Text Text)

The tags for the action.

newAction :: Action

Create a value of Action with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:parameters:Action', action_parameters - The action parameters, if applicable.

$sel:targets:Action', action_targets - The supported targets for the action.

$sel:id:Action', action_id - The ID of the action.

$sel:description:Action', action_description - The description for the action.

$sel:tags:Action', action_tags - The tags for the action.

action_parameters :: Lens' Action (Maybe (HashMap Text ActionParameter))

The action parameters, if applicable.

action_targets :: Lens' Action (Maybe (HashMap Text ActionTarget))

The supported targets for the action.

action_id :: Lens' Action (Maybe Text)

The ID of the action.

action_description :: Lens' Action (Maybe Text)

The description for the action.

action_tags :: Lens' Action (Maybe (HashMap Text Text))

The tags for the action.

ActionParameter

data ActionParameter

Describes a parameter for an action.

See: newActionParameter smart constructor.

Constructors

ActionParameter'

Fields

required :: Maybe Bool

Indicates whether the parameter is required.

description :: Maybe Text

The parameter description.

newActionParameter :: ActionParameter

Create a value of ActionParameter with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:required:ActionParameter', actionParameter_required - Indicates whether the parameter is required.

$sel:description:ActionParameter', actionParameter_description - The parameter description.

actionParameter_required :: Lens' ActionParameter (Maybe Bool)

Indicates whether the parameter is required.

actionParameter_description :: Lens' ActionParameter (Maybe Text)

The parameter description.

ActionSummary

data ActionSummary

Provides a summary of an action.

See: newActionSummary smart constructor.

Constructors

ActionSummary'

Fields

targets :: Maybe (HashMap Text ActionTarget)

The targets for the action.

id :: Maybe Text

The ID of the action.

description :: Maybe Text

The description for the action.

tags :: Maybe (HashMap Text Text)

The tags for the action.

newActionSummary :: ActionSummary

Create a value of ActionSummary with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:targets:ActionSummary', actionSummary_targets - The targets for the action.

$sel:id:ActionSummary', actionSummary_id - The ID of the action.

$sel:description:ActionSummary', actionSummary_description - The description for the action.

$sel:tags:ActionSummary', actionSummary_tags - The tags for the action.

actionSummary_targets :: Lens' ActionSummary (Maybe (HashMap Text ActionTarget))

The targets for the action.

actionSummary_id :: Lens' ActionSummary (Maybe Text)

The ID of the action.

actionSummary_description :: Lens' ActionSummary (Maybe Text)

The description for the action.

actionSummary_tags :: Lens' ActionSummary (Maybe (HashMap Text Text))

The tags for the action.

ActionTarget

data ActionTarget

Describes a target for an action.

See: newActionTarget smart constructor.

Constructors

ActionTarget'

Fields

resourceType :: Maybe Text

The resource type of the target.

newActionTarget :: ActionTarget

Create a value of ActionTarget with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:resourceType:ActionTarget', actionTarget_resourceType - The resource type of the target.

actionTarget_resourceType :: Lens' ActionTarget (Maybe Text)

The resource type of the target.

CreateExperimentTemplateActionInput

data CreateExperimentTemplateActionInput

Specifies an action for an experiment template.

See: newCreateExperimentTemplateActionInput smart constructor.

Constructors

CreateExperimentTemplateActionInput'

Fields

startAfter :: Maybe [Text]

The name of the action that must be completed before the current action starts. Omit this parameter to run the action at the start of the experiment.

parameters :: Maybe (HashMap Text Text)

The parameters for the action, if applicable.

targets :: Maybe (HashMap Text Text)

The targets for the action.

description :: Maybe Text

A description for the action.

actionId :: Text

The ID of the action.

newCreateExperimentTemplateActionInput

Arguments

:: Text
$sel:actionId:CreateExperimentTemplateActionInput'

-> CreateExperimentTemplateActionInput

Create a value of CreateExperimentTemplateActionInput with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:startAfter:CreateExperimentTemplateActionInput', createExperimentTemplateActionInput_startAfter - The name of the action that must be completed before the current action starts. Omit this parameter to run the action at the start of the experiment.

$sel:parameters:CreateExperimentTemplateActionInput', createExperimentTemplateActionInput_parameters - The parameters for the action, if applicable.

$sel:targets:CreateExperimentTemplateActionInput', createExperimentTemplateActionInput_targets - The targets for the action.

$sel:description:CreateExperimentTemplateActionInput', createExperimentTemplateActionInput_description - A description for the action.

$sel:actionId:CreateExperimentTemplateActionInput', createExperimentTemplateActionInput_actionId - The ID of the action.

createExperimentTemplateActionInput_startAfter :: Lens' CreateExperimentTemplateActionInput (Maybe [Text])

The name of the action that must be completed before the current action starts. Omit this parameter to run the action at the start of the experiment.

createExperimentTemplateActionInput_parameters :: Lens' CreateExperimentTemplateActionInput (Maybe (HashMap Text Text))

The parameters for the action, if applicable.

createExperimentTemplateActionInput_targets :: Lens' CreateExperimentTemplateActionInput (Maybe (HashMap Text Text))

The targets for the action.

createExperimentTemplateActionInput_description :: Lens' CreateExperimentTemplateActionInput (Maybe Text)

A description for the action.

createExperimentTemplateActionInput_actionId :: Lens' CreateExperimentTemplateActionInput Text

The ID of the action.

CreateExperimentTemplateStopConditionInput

data CreateExperimentTemplateStopConditionInput

Specifies a stop condition for an experiment template.

See: newCreateExperimentTemplateStopConditionInput smart constructor.

Constructors

CreateExperimentTemplateStopConditionInput'

Fields

value :: Maybe Text

The Amazon Resource Name (ARN) of the CloudWatch alarm. This is required if the source is a CloudWatch alarm.

source :: Text

The source for the stop condition. Specify aws:cloudwatch:alarm if the stop condition is defined by a CloudWatch alarm. Specify none if there is no stop condition.

newCreateExperimentTemplateStopConditionInput

Arguments

:: Text
$sel:source:CreateExperimentTemplateStopConditionInput'

-> CreateExperimentTemplateStopConditionInput

Create a value of CreateExperimentTemplateStopConditionInput with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:value:CreateExperimentTemplateStopConditionInput', createExperimentTemplateStopConditionInput_value - The Amazon Resource Name (ARN) of the CloudWatch alarm. This is required if the source is a CloudWatch alarm.

$sel:source:CreateExperimentTemplateStopConditionInput', createExperimentTemplateStopConditionInput_source - The source for the stop condition. Specify aws:cloudwatch:alarm if the stop condition is defined by a CloudWatch alarm. Specify none if there is no stop condition.

createExperimentTemplateStopConditionInput_value :: Lens' CreateExperimentTemplateStopConditionInput (Maybe Text)

The Amazon Resource Name (ARN) of the CloudWatch alarm. This is required if the source is a CloudWatch alarm.

createExperimentTemplateStopConditionInput_source :: Lens' CreateExperimentTemplateStopConditionInput Text

The source for the stop condition. Specify aws:cloudwatch:alarm if the stop condition is defined by a CloudWatch alarm. Specify none if there is no stop condition.

CreateExperimentTemplateTargetInput

data CreateExperimentTemplateTargetInput

Specifies a target for an experiment. You must specify at least one Amazon Resource Name (ARN) or at least one resource tag. You cannot specify both ARNs and tags.

See: newCreateExperimentTemplateTargetInput smart constructor.

Constructors

CreateExperimentTemplateTargetInput'

Fields

resourceTags :: Maybe (HashMap Text Text)

The tags for the target resources.

filters :: Maybe [ExperimentTemplateTargetInputFilter]

The filters to apply to identify target resources using specific attributes.

resourceArns :: Maybe [Text]

The Amazon Resource Names (ARNs) of the resources.

resourceType :: Text

The AWS resource type. The resource type must be supported for the specified action.

selectionMode :: Text

Scopes the identified resources to a specific count of the resources at random, or a percentage of the resources. All identified resources are included in the target.

ALL - Run the action on all identified targets. This is the default.

COUNT(n) - Run the action on the specified number of targets, chosen from the identified targets at random. For example, COUNT(1) selects one of the targets.

PERCENT(n) - Run the action on the specified percentage of targets, chosen from the identified targets at random. For example, PERCENT(25) selects 25% of the targets.

newCreateExperimentTemplateTargetInput

Arguments

:: Text
$sel:resourceType:CreateExperimentTemplateTargetInput'

-> Text
$sel:selectionMode:CreateExperimentTemplateTargetInput'

-> CreateExperimentTemplateTargetInput

Create a value of CreateExperimentTemplateTargetInput with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:resourceTags:CreateExperimentTemplateTargetInput', createExperimentTemplateTargetInput_resourceTags - The tags for the target resources.

$sel:filters:CreateExperimentTemplateTargetInput', createExperimentTemplateTargetInput_filters - The filters to apply to identify target resources using specific attributes.

$sel:resourceArns:CreateExperimentTemplateTargetInput', createExperimentTemplateTargetInput_resourceArns - The Amazon Resource Names (ARNs) of the resources.

$sel:resourceType:CreateExperimentTemplateTargetInput', createExperimentTemplateTargetInput_resourceType - The AWS resource type. The resource type must be supported for the specified action.

$sel:selectionMode:CreateExperimentTemplateTargetInput', createExperimentTemplateTargetInput_selectionMode - Scopes the identified resources to a specific count of the resources at random, or a percentage of the resources. All identified resources are included in the target.

ALL - Run the action on all identified targets. This is the default.

COUNT(n) - Run the action on the specified number of targets, chosen from the identified targets at random. For example, COUNT(1) selects one of the targets.

PERCENT(n) - Run the action on the specified percentage of targets, chosen from the identified targets at random. For example, PERCENT(25) selects 25% of the targets.

createExperimentTemplateTargetInput_resourceTags :: Lens' CreateExperimentTemplateTargetInput (Maybe (HashMap Text Text))

The tags for the target resources.

createExperimentTemplateTargetInput_filters :: Lens' CreateExperimentTemplateTargetInput (Maybe [ExperimentTemplateTargetInputFilter])

The filters to apply to identify target resources using specific attributes.

createExperimentTemplateTargetInput_resourceArns :: Lens' CreateExperimentTemplateTargetInput (Maybe [Text])

The Amazon Resource Names (ARNs) of the resources.

createExperimentTemplateTargetInput_resourceType :: Lens' CreateExperimentTemplateTargetInput Text

The AWS resource type. The resource type must be supported for the specified action.

createExperimentTemplateTargetInput_selectionMode :: Lens' CreateExperimentTemplateTargetInput Text

Scopes the identified resources to a specific count of the resources at random, or a percentage of the resources. All identified resources are included in the target.

ALL - Run the action on all identified targets. This is the default.

COUNT(n) - Run the action on the specified number of targets, chosen from the identified targets at random. For example, COUNT(1) selects one of the targets.

PERCENT(n) - Run the action on the specified percentage of targets, chosen from the identified targets at random. For example, PERCENT(25) selects 25% of the targets.

Experiment

data Experiment

Describes an experiment.

See: newExperiment smart constructor.

Constructors

Experiment'

Fields

creationTime :: Maybe POSIX

The time the experiment was created.

experimentTemplateId :: Maybe Text

The ID of the experiment template.

state :: Maybe ExperimentState

The state of the experiment.

actions :: Maybe (HashMap Text ExperimentAction)

The actions for the experiment.

startTime :: Maybe POSIX

The time that the experiment was started.

stopConditions :: Maybe [ExperimentStopCondition]

The stop conditions for the experiment.

endTime :: Maybe POSIX

The time that the experiment ended.

targets :: Maybe (HashMap Text ExperimentTarget)

The targets for the experiment.

id :: Maybe Text

The ID of the experiment.

tags :: Maybe (HashMap Text Text)

The tags for the experiment.

roleArn :: Maybe Text

The Amazon Resource Name (ARN) of an IAM role that grants the AWS FIS service permission to perform service actions on your behalf.

newExperiment :: Experiment

Create a value of Experiment with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:creationTime:Experiment', experiment_creationTime - The time the experiment was created.

$sel:experimentTemplateId:Experiment', experiment_experimentTemplateId - The ID of the experiment template.

$sel:state:Experiment', experiment_state - The state of the experiment.

$sel:actions:Experiment', experiment_actions - The actions for the experiment.

$sel:startTime:Experiment', experiment_startTime - The time that the experiment was started.

$sel:stopConditions:Experiment', experiment_stopConditions - The stop conditions for the experiment.

$sel:endTime:Experiment', experiment_endTime - The time that the experiment ended.

$sel:targets:Experiment', experiment_targets - The targets for the experiment.

$sel:id:Experiment', experiment_id - The ID of the experiment.

$sel:tags:Experiment', experiment_tags - The tags for the experiment.

$sel:roleArn:Experiment', experiment_roleArn - The Amazon Resource Name (ARN) of an IAM role that grants the AWS FIS service permission to perform service actions on your behalf.

experiment_creationTime :: Lens' Experiment (Maybe UTCTime)

The time the experiment was created.

experiment_experimentTemplateId :: Lens' Experiment (Maybe Text)

The ID of the experiment template.

experiment_state :: Lens' Experiment (Maybe ExperimentState)

The state of the experiment.

experiment_actions :: Lens' Experiment (Maybe (HashMap Text ExperimentAction))

The actions for the experiment.

experiment_startTime :: Lens' Experiment (Maybe UTCTime)

The time that the experiment was started.

experiment_stopConditions :: Lens' Experiment (Maybe [ExperimentStopCondition])

The stop conditions for the experiment.

experiment_endTime :: Lens' Experiment (Maybe UTCTime)

The time that the experiment ended.

experiment_targets :: Lens' Experiment (Maybe (HashMap Text ExperimentTarget))

The targets for the experiment.

experiment_id :: Lens' Experiment (Maybe Text)

The ID of the experiment.

experiment_tags :: Lens' Experiment (Maybe (HashMap Text Text))

The tags for the experiment.

experiment_roleArn :: Lens' Experiment (Maybe Text)

The Amazon Resource Name (ARN) of an IAM role that grants the AWS FIS service permission to perform service actions on your behalf.

ExperimentAction

data ExperimentAction

Describes the action for an experiment.

See: newExperimentAction smart constructor.

Constructors

ExperimentAction'

Fields

startAfter :: Maybe [Text]

The name of the action that must be completed before this action starts.

state :: Maybe ExperimentActionState

The state of the action.

actionId :: Maybe Text

The ID of the action.

parameters :: Maybe (HashMap Text Text)

The parameters for the action.

targets :: Maybe (HashMap Text Text)

The targets for the action.

description :: Maybe Text

The description for the action.

newExperimentAction :: ExperimentAction

Create a value of ExperimentAction with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:startAfter:ExperimentAction', experimentAction_startAfter - The name of the action that must be completed before this action starts.

$sel:state:ExperimentAction', experimentAction_state - The state of the action.

$sel:actionId:ExperimentAction', experimentAction_actionId - The ID of the action.

$sel:parameters:ExperimentAction', experimentAction_parameters - The parameters for the action.

$sel:targets:ExperimentAction', experimentAction_targets - The targets for the action.

$sel:description:ExperimentAction', experimentAction_description - The description for the action.

experimentAction_startAfter :: Lens' ExperimentAction (Maybe [Text])

The name of the action that must be completed before this action starts.

experimentAction_state :: Lens' ExperimentAction (Maybe ExperimentActionState)

The state of the action.

experimentAction_actionId :: Lens' ExperimentAction (Maybe Text)

The ID of the action.

experimentAction_parameters :: Lens' ExperimentAction (Maybe (HashMap Text Text))

The parameters for the action.

experimentAction_targets :: Lens' ExperimentAction (Maybe (HashMap Text Text))

The targets for the action.

experimentAction_description :: Lens' ExperimentAction (Maybe Text)

The description for the action.

ExperimentActionState

data ExperimentActionState

Describes the state of an action.

See: newExperimentActionState smart constructor.

Constructors

ExperimentActionState'

Fields

status :: Maybe ExperimentActionStatus

The state of the action.

reason :: Maybe Text

The reason for the state.

newExperimentActionState :: ExperimentActionState

Create a value of ExperimentActionState with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:status:ExperimentActionState', experimentActionState_status - The state of the action.

$sel:reason:ExperimentActionState', experimentActionState_reason - The reason for the state.

experimentActionState_status :: Lens' ExperimentActionState (Maybe ExperimentActionStatus)

The state of the action.

experimentActionState_reason :: Lens' ExperimentActionState (Maybe Text)

The reason for the state.

ExperimentState

data ExperimentState

Describes the state of an experiment.

See: newExperimentState smart constructor.

Constructors

ExperimentState'

Fields

status :: Maybe ExperimentStatus

The state of the experiment.

reason :: Maybe Text

The reason for the state.

newExperimentState :: ExperimentState

Create a value of ExperimentState with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:status:ExperimentState', experimentState_status - The state of the experiment.

$sel:reason:ExperimentState', experimentState_reason - The reason for the state.

experimentState_status :: Lens' ExperimentState (Maybe ExperimentStatus)

The state of the experiment.

experimentState_reason :: Lens' ExperimentState (Maybe Text)

The reason for the state.

ExperimentStopCondition

data ExperimentStopCondition

Describes the stop condition for an experiment.

See: newExperimentStopCondition smart constructor.

Constructors

ExperimentStopCondition'

Fields

value :: Maybe Text

The Amazon Resource Name (ARN) of the CloudWatch alarm, if applicable.

source :: Maybe Text

The source for the stop condition.

newExperimentStopCondition :: ExperimentStopCondition

Create a value of ExperimentStopCondition with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:value:ExperimentStopCondition', experimentStopCondition_value - The Amazon Resource Name (ARN) of the CloudWatch alarm, if applicable.

$sel:source:ExperimentStopCondition', experimentStopCondition_source - The source for the stop condition.

experimentStopCondition_value :: Lens' ExperimentStopCondition (Maybe Text)

The Amazon Resource Name (ARN) of the CloudWatch alarm, if applicable.

experimentStopCondition_source :: Lens' ExperimentStopCondition (Maybe Text)

The source for the stop condition.

ExperimentSummary

data ExperimentSummary

Provides a summary of an experiment.

See: newExperimentSummary smart constructor.

Constructors

ExperimentSummary'

Fields

creationTime :: Maybe POSIX

The time that the experiment was created.

experimentTemplateId :: Maybe Text

The ID of the experiment template.

state :: Maybe ExperimentState

The state of the experiment.

id :: Maybe Text

The ID of the experiment.

tags :: Maybe (HashMap Text Text)

The tags for the experiment.

newExperimentSummary :: ExperimentSummary

Create a value of ExperimentSummary with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:creationTime:ExperimentSummary', experimentSummary_creationTime - The time that the experiment was created.

$sel:experimentTemplateId:ExperimentSummary', experimentSummary_experimentTemplateId - The ID of the experiment template.

$sel:state:ExperimentSummary', experimentSummary_state - The state of the experiment.

$sel:id:ExperimentSummary', experimentSummary_id - The ID of the experiment.

$sel:tags:ExperimentSummary', experimentSummary_tags - The tags for the experiment.

experimentSummary_creationTime :: Lens' ExperimentSummary (Maybe UTCTime)

The time that the experiment was created.

experimentSummary_experimentTemplateId :: Lens' ExperimentSummary (Maybe Text)

The ID of the experiment template.

experimentSummary_state :: Lens' ExperimentSummary (Maybe ExperimentState)

The state of the experiment.

experimentSummary_id :: Lens' ExperimentSummary (Maybe Text)

The ID of the experiment.

experimentSummary_tags :: Lens' ExperimentSummary (Maybe (HashMap Text Text))

The tags for the experiment.

ExperimentTarget

data ExperimentTarget

Describes a target for an experiment.

See: newExperimentTarget smart constructor.

Constructors

ExperimentTarget'

Fields

resourceType :: Maybe Text

The resource type.

resourceTags :: Maybe (HashMap Text Text)

The tags for the target resources.

filters :: Maybe [ExperimentTargetFilter]

The filters to apply to identify target resources using specific attributes.

resourceArns :: Maybe [Text]

The Amazon Resource Names (ARNs) of the resources.

selectionMode :: Maybe Text

Scopes the identified resources to a specific count or percentage.

newExperimentTarget :: ExperimentTarget

Create a value of ExperimentTarget with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:resourceType:ExperimentTarget', experimentTarget_resourceType - The resource type.

$sel:resourceTags:ExperimentTarget', experimentTarget_resourceTags - The tags for the target resources.

$sel:filters:ExperimentTarget', experimentTarget_filters - The filters to apply to identify target resources using specific attributes.

$sel:resourceArns:ExperimentTarget', experimentTarget_resourceArns - The Amazon Resource Names (ARNs) of the resources.

$sel:selectionMode:ExperimentTarget', experimentTarget_selectionMode - Scopes the identified resources to a specific count or percentage.

experimentTarget_resourceType :: Lens' ExperimentTarget (Maybe Text)

The resource type.

experimentTarget_resourceTags :: Lens' ExperimentTarget (Maybe (HashMap Text Text))

The tags for the target resources.

experimentTarget_filters :: Lens' ExperimentTarget (Maybe [ExperimentTargetFilter])

The filters to apply to identify target resources using specific attributes.

experimentTarget_resourceArns :: Lens' ExperimentTarget (Maybe [Text])

The Amazon Resource Names (ARNs) of the resources.

experimentTarget_selectionMode :: Lens' ExperimentTarget (Maybe Text)

Scopes the identified resources to a specific count or percentage.

ExperimentTargetFilter

data ExperimentTargetFilter

Describes a filter used for the target resources in an experiment.

See: newExperimentTargetFilter smart constructor.

Constructors

ExperimentTargetFilter'

Fields

path :: Maybe Text

The attribute path for the filter.

values :: Maybe [Text]

The attribute values for the filter.

newExperimentTargetFilter :: ExperimentTargetFilter

Create a value of ExperimentTargetFilter with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:path:ExperimentTargetFilter', experimentTargetFilter_path - The attribute path for the filter.

$sel:values:ExperimentTargetFilter', experimentTargetFilter_values - The attribute values for the filter.

experimentTargetFilter_path :: Lens' ExperimentTargetFilter (Maybe Text)

The attribute path for the filter.

experimentTargetFilter_values :: Lens' ExperimentTargetFilter (Maybe [Text])

The attribute values for the filter.

ExperimentTemplate

data ExperimentTemplate

Describes an experiment template.

See: newExperimentTemplate smart constructor.

Constructors

ExperimentTemplate'

Fields

creationTime :: Maybe POSIX

The time the experiment template was created.

actions :: Maybe (HashMap Text ExperimentTemplateAction)

The actions for the experiment.

stopConditions :: Maybe [ExperimentTemplateStopCondition]

The stop conditions for the experiment.

targets :: Maybe (HashMap Text ExperimentTemplateTarget)

The targets for the experiment.

id :: Maybe Text

The ID of the experiment template.

lastUpdateTime :: Maybe POSIX

The time the experiment template was last updated.

description :: Maybe Text

The description for the experiment template.

tags :: Maybe (HashMap Text Text)

The tags for the experiment template.

roleArn :: Maybe Text

The Amazon Resource Name (ARN) of an IAM role.

newExperimentTemplate :: ExperimentTemplate

Create a value of ExperimentTemplate with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:creationTime:ExperimentTemplate', experimentTemplate_creationTime - The time the experiment template was created.

$sel:actions:ExperimentTemplate', experimentTemplate_actions - The actions for the experiment.

$sel:stopConditions:ExperimentTemplate', experimentTemplate_stopConditions - The stop conditions for the experiment.

$sel:targets:ExperimentTemplate', experimentTemplate_targets - The targets for the experiment.

$sel:id:ExperimentTemplate', experimentTemplate_id - The ID of the experiment template.

$sel:lastUpdateTime:ExperimentTemplate', experimentTemplate_lastUpdateTime - The time the experiment template was last updated.

$sel:description:ExperimentTemplate', experimentTemplate_description - The description for the experiment template.

$sel:tags:ExperimentTemplate', experimentTemplate_tags - The tags for the experiment template.

$sel:roleArn:ExperimentTemplate', experimentTemplate_roleArn - The Amazon Resource Name (ARN) of an IAM role.

experimentTemplate_creationTime :: Lens' ExperimentTemplate (Maybe UTCTime)

The time the experiment template was created.

experimentTemplate_actions :: Lens' ExperimentTemplate (Maybe (HashMap Text ExperimentTemplateAction))

The actions for the experiment.

experimentTemplate_stopConditions :: Lens' ExperimentTemplate (Maybe [ExperimentTemplateStopCondition])

The stop conditions for the experiment.

experimentTemplate_targets :: Lens' ExperimentTemplate (Maybe (HashMap Text ExperimentTemplateTarget))

The targets for the experiment.

experimentTemplate_id :: Lens' ExperimentTemplate (Maybe Text)

The ID of the experiment template.

experimentTemplate_lastUpdateTime :: Lens' ExperimentTemplate (Maybe UTCTime)

The time the experiment template was last updated.

experimentTemplate_description :: Lens' ExperimentTemplate (Maybe Text)

The description for the experiment template.

experimentTemplate_tags :: Lens' ExperimentTemplate (Maybe (HashMap Text Text))

The tags for the experiment template.

experimentTemplate_roleArn :: Lens' ExperimentTemplate (Maybe Text)

The Amazon Resource Name (ARN) of an IAM role.

ExperimentTemplateAction

data ExperimentTemplateAction

Describes an action for an experiment template.

See: newExperimentTemplateAction smart constructor.

Constructors

ExperimentTemplateAction'

Fields

startAfter :: Maybe [Text]

The name of the action that must be completed before the current action starts.

actionId :: Maybe Text

The ID of the action.

parameters :: Maybe (HashMap Text Text)

The parameters for the action.

targets :: Maybe (HashMap Text Text)

The targets for the action.

description :: Maybe Text

A description for the action.

newExperimentTemplateAction :: ExperimentTemplateAction

Create a value of ExperimentTemplateAction with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:startAfter:ExperimentTemplateAction', experimentTemplateAction_startAfter - The name of the action that must be completed before the current action starts.

$sel:actionId:ExperimentTemplateAction', experimentTemplateAction_actionId - The ID of the action.

$sel:parameters:ExperimentTemplateAction', experimentTemplateAction_parameters - The parameters for the action.

$sel:targets:ExperimentTemplateAction', experimentTemplateAction_targets - The targets for the action.

$sel:description:ExperimentTemplateAction', experimentTemplateAction_description - A description for the action.

experimentTemplateAction_startAfter :: Lens' ExperimentTemplateAction (Maybe [Text])

The name of the action that must be completed before the current action starts.

experimentTemplateAction_actionId :: Lens' ExperimentTemplateAction (Maybe Text)

The ID of the action.

experimentTemplateAction_parameters :: Lens' ExperimentTemplateAction (Maybe (HashMap Text Text))

The parameters for the action.

experimentTemplateAction_targets :: Lens' ExperimentTemplateAction (Maybe (HashMap Text Text))

The targets for the action.

experimentTemplateAction_description :: Lens' ExperimentTemplateAction (Maybe Text)

A description for the action.

ExperimentTemplateStopCondition

data ExperimentTemplateStopCondition

Describes a stop condition for an experiment template.

See: newExperimentTemplateStopCondition smart constructor.

Constructors

ExperimentTemplateStopCondition'

Fields

value :: Maybe Text

The Amazon Resource Name (ARN) of the CloudWatch alarm, if applicable.

source :: Maybe Text

The source for the stop condition.

newExperimentTemplateStopCondition :: ExperimentTemplateStopCondition

Create a value of ExperimentTemplateStopCondition with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:value:ExperimentTemplateStopCondition', experimentTemplateStopCondition_value - The Amazon Resource Name (ARN) of the CloudWatch alarm, if applicable.

$sel:source:ExperimentTemplateStopCondition', experimentTemplateStopCondition_source - The source for the stop condition.

experimentTemplateStopCondition_value :: Lens' ExperimentTemplateStopCondition (Maybe Text)

The Amazon Resource Name (ARN) of the CloudWatch alarm, if applicable.

experimentTemplateStopCondition_source :: Lens' ExperimentTemplateStopCondition (Maybe Text)

The source for the stop condition.

ExperimentTemplateSummary

data ExperimentTemplateSummary

Provides a summary of an experiment template.

See: newExperimentTemplateSummary smart constructor.

Constructors

ExperimentTemplateSummary'

Fields

creationTime :: Maybe POSIX

The time that the experiment template was created.

id :: Maybe Text

The ID of the experiment template.

lastUpdateTime :: Maybe POSIX

The time that the experiment template was last updated.

description :: Maybe Text

The description of the experiment template.

tags :: Maybe (HashMap Text Text)

The tags for the experiment template.

newExperimentTemplateSummary :: ExperimentTemplateSummary

Create a value of ExperimentTemplateSummary with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:creationTime:ExperimentTemplateSummary', experimentTemplateSummary_creationTime - The time that the experiment template was created.

$sel:id:ExperimentTemplateSummary', experimentTemplateSummary_id - The ID of the experiment template.

$sel:lastUpdateTime:ExperimentTemplateSummary', experimentTemplateSummary_lastUpdateTime - The time that the experiment template was last updated.

$sel:description:ExperimentTemplateSummary', experimentTemplateSummary_description - The description of the experiment template.

$sel:tags:ExperimentTemplateSummary', experimentTemplateSummary_tags - The tags for the experiment template.

experimentTemplateSummary_creationTime :: Lens' ExperimentTemplateSummary (Maybe UTCTime)

The time that the experiment template was created.

experimentTemplateSummary_id :: Lens' ExperimentTemplateSummary (Maybe Text)

The ID of the experiment template.

experimentTemplateSummary_lastUpdateTime :: Lens' ExperimentTemplateSummary (Maybe UTCTime)

The time that the experiment template was last updated.

experimentTemplateSummary_description :: Lens' ExperimentTemplateSummary (Maybe Text)

The description of the experiment template.

experimentTemplateSummary_tags :: Lens' ExperimentTemplateSummary (Maybe (HashMap Text Text))

The tags for the experiment template.

ExperimentTemplateTarget

data ExperimentTemplateTarget

Describes a target for an experiment template.

See: newExperimentTemplateTarget smart constructor.

Constructors

ExperimentTemplateTarget'

Fields

resourceType :: Maybe Text

The resource type.

resourceTags :: Maybe (HashMap Text Text)

The tags for the target resources.

filters :: Maybe [ExperimentTemplateTargetFilter]

The filters to apply to identify target resources using specific attributes.

resourceArns :: Maybe [Text]

The Amazon Resource Names (ARNs) of the targets.

selectionMode :: Maybe Text

Scopes the identified resources to a specific count or percentage.

newExperimentTemplateTarget :: ExperimentTemplateTarget

Create a value of ExperimentTemplateTarget with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:resourceType:ExperimentTemplateTarget', experimentTemplateTarget_resourceType - The resource type.

$sel:resourceTags:ExperimentTemplateTarget', experimentTemplateTarget_resourceTags - The tags for the target resources.

$sel:filters:ExperimentTemplateTarget', experimentTemplateTarget_filters - The filters to apply to identify target resources using specific attributes.

$sel:resourceArns:ExperimentTemplateTarget', experimentTemplateTarget_resourceArns - The Amazon Resource Names (ARNs) of the targets.

$sel:selectionMode:ExperimentTemplateTarget', experimentTemplateTarget_selectionMode - Scopes the identified resources to a specific count or percentage.

experimentTemplateTarget_resourceType :: Lens' ExperimentTemplateTarget (Maybe Text)

The resource type.

experimentTemplateTarget_resourceTags :: Lens' ExperimentTemplateTarget (Maybe (HashMap Text Text))

The tags for the target resources.

experimentTemplateTarget_filters :: Lens' ExperimentTemplateTarget (Maybe [ExperimentTemplateTargetFilter])

The filters to apply to identify target resources using specific attributes.

experimentTemplateTarget_resourceArns :: Lens' ExperimentTemplateTarget (Maybe [Text])

The Amazon Resource Names (ARNs) of the targets.

experimentTemplateTarget_selectionMode :: Lens' ExperimentTemplateTarget (Maybe Text)

Scopes the identified resources to a specific count or percentage.

ExperimentTemplateTargetFilter

data ExperimentTemplateTargetFilter

Describes a filter used for the target resources in an experiment template.

See: newExperimentTemplateTargetFilter smart constructor.

Constructors

ExperimentTemplateTargetFilter'

Fields

path :: Maybe Text

The attribute path for the filter.

values :: Maybe [Text]

The attribute values for the filter.

newExperimentTemplateTargetFilter :: ExperimentTemplateTargetFilter

Create a value of ExperimentTemplateTargetFilter with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:path:ExperimentTemplateTargetFilter', experimentTemplateTargetFilter_path - The attribute path for the filter.

$sel:values:ExperimentTemplateTargetFilter', experimentTemplateTargetFilter_values - The attribute values for the filter.

experimentTemplateTargetFilter_path :: Lens' ExperimentTemplateTargetFilter (Maybe Text)

The attribute path for the filter.

experimentTemplateTargetFilter_values :: Lens' ExperimentTemplateTargetFilter (Maybe [Text])

The attribute values for the filter.

ExperimentTemplateTargetInputFilter

data ExperimentTemplateTargetInputFilter

Describes a filter used for the target resource input in an experiment template.

See: newExperimentTemplateTargetInputFilter smart constructor.

Constructors

ExperimentTemplateTargetInputFilter'

Fields

path :: Text

The attribute path for the filter.

values :: [Text]

The attribute values for the filter.

newExperimentTemplateTargetInputFilter

Arguments

:: Text
$sel:path:ExperimentTemplateTargetInputFilter'

-> ExperimentTemplateTargetInputFilter

Create a value of ExperimentTemplateTargetInputFilter with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:path:ExperimentTemplateTargetInputFilter', experimentTemplateTargetInputFilter_path - The attribute path for the filter.

$sel:values:ExperimentTemplateTargetInputFilter', experimentTemplateTargetInputFilter_values - The attribute values for the filter.

experimentTemplateTargetInputFilter_path :: Lens' ExperimentTemplateTargetInputFilter Text

The attribute path for the filter.

experimentTemplateTargetInputFilter_values :: Lens' ExperimentTemplateTargetInputFilter [Text]

The attribute values for the filter.

UpdateExperimentTemplateActionInputItem

data UpdateExperimentTemplateActionInputItem

Specifies an action for an experiment template.

See: newUpdateExperimentTemplateActionInputItem smart constructor.

Constructors

UpdateExperimentTemplateActionInputItem'

Fields

startAfter :: Maybe [Text]

The name of the action that must be completed before the current action starts. Omit this parameter to run the action at the start of the experiment.

actionId :: Maybe Text

The ID of the action.

parameters :: Maybe (HashMap Text Text)

The parameters for the action, if applicable.

targets :: Maybe (HashMap Text Text)

The targets for the action.

description :: Maybe Text

A description for the action.

newUpdateExperimentTemplateActionInputItem :: UpdateExperimentTemplateActionInputItem

Create a value of UpdateExperimentTemplateActionInputItem with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:startAfter:UpdateExperimentTemplateActionInputItem', updateExperimentTemplateActionInputItem_startAfter - The name of the action that must be completed before the current action starts. Omit this parameter to run the action at the start of the experiment.

$sel:actionId:UpdateExperimentTemplateActionInputItem', updateExperimentTemplateActionInputItem_actionId - The ID of the action.

$sel:parameters:UpdateExperimentTemplateActionInputItem', updateExperimentTemplateActionInputItem_parameters - The parameters for the action, if applicable.

$sel:targets:UpdateExperimentTemplateActionInputItem', updateExperimentTemplateActionInputItem_targets - The targets for the action.

$sel:description:UpdateExperimentTemplateActionInputItem', updateExperimentTemplateActionInputItem_description - A description for the action.

updateExperimentTemplateActionInputItem_startAfter :: Lens' UpdateExperimentTemplateActionInputItem (Maybe [Text])

The name of the action that must be completed before the current action starts. Omit this parameter to run the action at the start of the experiment.

updateExperimentTemplateActionInputItem_actionId :: Lens' UpdateExperimentTemplateActionInputItem (Maybe Text)

The ID of the action.

updateExperimentTemplateActionInputItem_parameters :: Lens' UpdateExperimentTemplateActionInputItem (Maybe (HashMap Text Text))

The parameters for the action, if applicable.

updateExperimentTemplateActionInputItem_targets :: Lens' UpdateExperimentTemplateActionInputItem (Maybe (HashMap Text Text))

The targets for the action.

updateExperimentTemplateActionInputItem_description :: Lens' UpdateExperimentTemplateActionInputItem (Maybe Text)

A description for the action.

UpdateExperimentTemplateStopConditionInput

data UpdateExperimentTemplateStopConditionInput

Specifies a stop condition for an experiment. You can define a stop condition as a CloudWatch alarm.

See: newUpdateExperimentTemplateStopConditionInput smart constructor.

Constructors

UpdateExperimentTemplateStopConditionInput'

Fields

value :: Maybe Text

The Amazon Resource Name (ARN) of the CloudWatch alarm.

source :: Text

The source for the stop condition. Specify aws:cloudwatch:alarm if the stop condition is defined by a CloudWatch alarm. Specify none if there is no stop condition.

newUpdateExperimentTemplateStopConditionInput

Arguments

:: Text
$sel:source:UpdateExperimentTemplateStopConditionInput'

-> UpdateExperimentTemplateStopConditionInput

Create a value of UpdateExperimentTemplateStopConditionInput with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:value:UpdateExperimentTemplateStopConditionInput', updateExperimentTemplateStopConditionInput_value - The Amazon Resource Name (ARN) of the CloudWatch alarm.

$sel:source:UpdateExperimentTemplateStopConditionInput', updateExperimentTemplateStopConditionInput_source - The source for the stop condition. Specify aws:cloudwatch:alarm if the stop condition is defined by a CloudWatch alarm. Specify none if there is no stop condition.

updateExperimentTemplateStopConditionInput_value :: Lens' UpdateExperimentTemplateStopConditionInput (Maybe Text)

The Amazon Resource Name (ARN) of the CloudWatch alarm.

updateExperimentTemplateStopConditionInput_source :: Lens' UpdateExperimentTemplateStopConditionInput Text

The source for the stop condition. Specify aws:cloudwatch:alarm if the stop condition is defined by a CloudWatch alarm. Specify none if there is no stop condition.

UpdateExperimentTemplateTargetInput

data UpdateExperimentTemplateTargetInput

Specifies a target for an experiment. You must specify at least one Amazon Resource Name (ARN) or at least one resource tag. You cannot specify both.

See: newUpdateExperimentTemplateTargetInput smart constructor.

Constructors

UpdateExperimentTemplateTargetInput'

Fields

resourceTags :: Maybe (HashMap Text Text)

The tags for the target resources.

filters :: Maybe [ExperimentTemplateTargetInputFilter]

The filters to apply to identify target resources using specific attributes.

resourceArns :: Maybe [Text]

The Amazon Resource Names (ARNs) of the targets.

resourceType :: Text

The AWS resource type. The resource type must be supported for the specified action.

selectionMode :: Text

Scopes the identified resources to a specific count or percentage.

newUpdateExperimentTemplateTargetInput

Arguments

:: Text
$sel:resourceType:UpdateExperimentTemplateTargetInput'

-> Text
$sel:selectionMode:UpdateExperimentTemplateTargetInput'

-> UpdateExperimentTemplateTargetInput

Create a value of UpdateExperimentTemplateTargetInput with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:resourceTags:UpdateExperimentTemplateTargetInput', updateExperimentTemplateTargetInput_resourceTags - The tags for the target resources.

$sel:filters:UpdateExperimentTemplateTargetInput', updateExperimentTemplateTargetInput_filters - The filters to apply to identify target resources using specific attributes.

$sel:resourceArns:UpdateExperimentTemplateTargetInput', updateExperimentTemplateTargetInput_resourceArns - The Amazon Resource Names (ARNs) of the targets.

$sel:resourceType:UpdateExperimentTemplateTargetInput', updateExperimentTemplateTargetInput_resourceType - The AWS resource type. The resource type must be supported for the specified action.

$sel:selectionMode:UpdateExperimentTemplateTargetInput', updateExperimentTemplateTargetInput_selectionMode - Scopes the identified resources to a specific count or percentage.

updateExperimentTemplateTargetInput_resourceTags :: Lens' UpdateExperimentTemplateTargetInput (Maybe (HashMap Text Text))

The tags for the target resources.

updateExperimentTemplateTargetInput_filters :: Lens' UpdateExperimentTemplateTargetInput (Maybe [ExperimentTemplateTargetInputFilter])

The filters to apply to identify target resources using specific attributes.

updateExperimentTemplateTargetInput_resourceArns :: Lens' UpdateExperimentTemplateTargetInput (Maybe [Text])

The Amazon Resource Names (ARNs) of the targets.

updateExperimentTemplateTargetInput_resourceType :: Lens' UpdateExperimentTemplateTargetInput Text

The AWS resource type. The resource type must be supported for the specified action.

updateExperimentTemplateTargetInput_selectionMode :: Lens' UpdateExperimentTemplateTargetInput Text

Scopes the identified resources to a specific count or percentage.