Amazonka.FIS.Types.Experiment

Description

Documentation

data Experiment

Describes an experiment.

See: newExperiment smart constructor.

Constructors

Experiment'

Fields

creationTime :: Maybe POSIX

The time the experiment was created.

experimentTemplateId :: Maybe Text

The ID of the experiment template.

state :: Maybe ExperimentState

The state of the experiment.

actions :: Maybe (HashMap Text ExperimentAction)

The actions for the experiment.

startTime :: Maybe POSIX

The time that the experiment was started.

stopConditions :: Maybe [ExperimentStopCondition]

The stop conditions for the experiment.

endTime :: Maybe POSIX

The time that the experiment ended.

targets :: Maybe (HashMap Text ExperimentTarget)

The targets for the experiment.

id :: Maybe Text

The ID of the experiment.

tags :: Maybe (HashMap Text Text)

The tags for the experiment.

roleArn :: Maybe Text

The Amazon Resource Name (ARN) of an IAM role that grants the AWS FIS service permission to perform service actions on your behalf.

newExperiment :: Experiment

Create a value of Experiment with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:creationTime:Experiment', experiment_creationTime - The time the experiment was created.

$sel:experimentTemplateId:Experiment', experiment_experimentTemplateId - The ID of the experiment template.

$sel:state:Experiment', experiment_state - The state of the experiment.

$sel:actions:Experiment', experiment_actions - The actions for the experiment.

$sel:startTime:Experiment', experiment_startTime - The time that the experiment was started.

$sel:stopConditions:Experiment', experiment_stopConditions - The stop conditions for the experiment.

$sel:endTime:Experiment', experiment_endTime - The time that the experiment ended.

$sel:targets:Experiment', experiment_targets - The targets for the experiment.

$sel:id:Experiment', experiment_id - The ID of the experiment.

$sel:tags:Experiment', experiment_tags - The tags for the experiment.

$sel:roleArn:Experiment', experiment_roleArn - The Amazon Resource Name (ARN) of an IAM role that grants the AWS FIS service permission to perform service actions on your behalf.

experiment_creationTime :: Lens' Experiment (Maybe UTCTime)

The time the experiment was created.

experiment_experimentTemplateId :: Lens' Experiment (Maybe Text)

The ID of the experiment template.

experiment_state :: Lens' Experiment (Maybe ExperimentState)

The state of the experiment.

experiment_actions :: Lens' Experiment (Maybe (HashMap Text ExperimentAction))

The actions for the experiment.

experiment_startTime :: Lens' Experiment (Maybe UTCTime)

The time that the experiment was started.

experiment_stopConditions :: Lens' Experiment (Maybe [ExperimentStopCondition])

The stop conditions for the experiment.

experiment_endTime :: Lens' Experiment (Maybe UTCTime)

The time that the experiment ended.

experiment_targets :: Lens' Experiment (Maybe (HashMap Text ExperimentTarget))

The targets for the experiment.

experiment_id :: Lens' Experiment (Maybe Text)

The ID of the experiment.

experiment_tags :: Lens' Experiment (Maybe (HashMap Text Text))

The tags for the experiment.

experiment_roleArn :: Lens' Experiment (Maybe Text)

The Amazon Resource Name (ARN) of an IAM role that grants the AWS FIS service permission to perform service actions on your behalf.