Amazonka.FIS.Types.ExperimentTemplateTarget

Description

Documentation

data ExperimentTemplateTarget

Describes a target for an experiment template.

See: newExperimentTemplateTarget smart constructor.

Constructors

ExperimentTemplateTarget'

Fields

resourceType :: Maybe Text

The resource type.

resourceTags :: Maybe (HashMap Text Text)

The tags for the target resources.

filters :: Maybe [ExperimentTemplateTargetFilter]

The filters to apply to identify target resources using specific attributes.

resourceArns :: Maybe [Text]

The Amazon Resource Names (ARNs) of the targets.

selectionMode :: Maybe Text

Scopes the identified resources to a specific count or percentage.

newExperimentTemplateTarget :: ExperimentTemplateTarget

Create a value of ExperimentTemplateTarget with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:resourceType:ExperimentTemplateTarget', experimentTemplateTarget_resourceType - The resource type.

$sel:resourceTags:ExperimentTemplateTarget', experimentTemplateTarget_resourceTags - The tags for the target resources.

$sel:filters:ExperimentTemplateTarget', experimentTemplateTarget_filters - The filters to apply to identify target resources using specific attributes.

$sel:resourceArns:ExperimentTemplateTarget', experimentTemplateTarget_resourceArns - The Amazon Resource Names (ARNs) of the targets.

$sel:selectionMode:ExperimentTemplateTarget', experimentTemplateTarget_selectionMode - Scopes the identified resources to a specific count or percentage.

experimentTemplateTarget_resourceType :: Lens' ExperimentTemplateTarget (Maybe Text)

The resource type.

experimentTemplateTarget_resourceTags :: Lens' ExperimentTemplateTarget (Maybe (HashMap Text Text))

The tags for the target resources.

experimentTemplateTarget_filters :: Lens' ExperimentTemplateTarget (Maybe [ExperimentTemplateTargetFilter])

The filters to apply to identify target resources using specific attributes.

experimentTemplateTarget_resourceArns :: Lens' ExperimentTemplateTarget (Maybe [Text])

The Amazon Resource Names (ARNs) of the targets.

experimentTemplateTarget_selectionMode :: Lens' ExperimentTemplateTarget (Maybe Text)

Scopes the identified resources to a specific count or percentage.