Amazonka.FIS.UpdateExperimentTemplate

Description

Updates the specified experiment template.

Creating a Request

data UpdateExperimentTemplate

See: newUpdateExperimentTemplate smart constructor.

Constructors

UpdateExperimentTemplate'

Fields

actions :: Maybe (HashMap Text UpdateExperimentTemplateActionInputItem)

The actions for the experiment.

stopConditions :: Maybe [UpdateExperimentTemplateStopConditionInput]

The stop conditions for the experiment.

targets :: Maybe (HashMap Text UpdateExperimentTemplateTargetInput)

The targets for the experiment.

description :: Maybe Text

A description for the template.

roleArn :: Maybe Text

The Amazon Resource Name (ARN) of an IAM role that grants the AWS FIS service permission to perform service actions on your behalf.

id :: Text

The ID of the experiment template.

newUpdateExperimentTemplate

Arguments

:: Text
$sel:id:UpdateExperimentTemplate'

-> UpdateExperimentTemplate

Create a value of UpdateExperimentTemplate with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:actions:UpdateExperimentTemplate', updateExperimentTemplate_actions - The actions for the experiment.

$sel:stopConditions:UpdateExperimentTemplate', updateExperimentTemplate_stopConditions - The stop conditions for the experiment.

$sel:targets:UpdateExperimentTemplate', updateExperimentTemplate_targets - The targets for the experiment.

$sel:description:UpdateExperimentTemplate', updateExperimentTemplate_description - A description for the template.

$sel:roleArn:UpdateExperimentTemplate', updateExperimentTemplate_roleArn - The Amazon Resource Name (ARN) of an IAM role that grants the AWS FIS service permission to perform service actions on your behalf.

$sel:id:UpdateExperimentTemplate', updateExperimentTemplate_id - The ID of the experiment template.

Request Lenses

updateExperimentTemplate_actions :: Lens' UpdateExperimentTemplate (Maybe (HashMap Text UpdateExperimentTemplateActionInputItem))

The actions for the experiment.

updateExperimentTemplate_stopConditions :: Lens' UpdateExperimentTemplate (Maybe [UpdateExperimentTemplateStopConditionInput])

The stop conditions for the experiment.

updateExperimentTemplate_targets :: Lens' UpdateExperimentTemplate (Maybe (HashMap Text UpdateExperimentTemplateTargetInput))

The targets for the experiment.

updateExperimentTemplate_description :: Lens' UpdateExperimentTemplate (Maybe Text)

A description for the template.

updateExperimentTemplate_roleArn :: Lens' UpdateExperimentTemplate (Maybe Text)

The Amazon Resource Name (ARN) of an IAM role that grants the AWS FIS service permission to perform service actions on your behalf.

updateExperimentTemplate_id :: Lens' UpdateExperimentTemplate Text

The ID of the experiment template.

Destructuring the Response

data UpdateExperimentTemplateResponse

See: newUpdateExperimentTemplateResponse smart constructor.

Constructors

UpdateExperimentTemplateResponse'

Fields

experimentTemplate :: Maybe ExperimentTemplate

Information about the experiment template.

httpStatus :: Int

The response's http status code.

newUpdateExperimentTemplateResponse

Arguments

:: Int
$sel:httpStatus:UpdateExperimentTemplateResponse'

-> UpdateExperimentTemplateResponse

Create a value of UpdateExperimentTemplateResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:experimentTemplate:UpdateExperimentTemplateResponse', updateExperimentTemplateResponse_experimentTemplate - Information about the experiment template.

$sel:httpStatus:UpdateExperimentTemplateResponse', updateExperimentTemplateResponse_httpStatus - The response's http status code.

Response Lenses

updateExperimentTemplateResponse_experimentTemplate :: Lens' UpdateExperimentTemplateResponse (Maybe ExperimentTemplate)

Information about the experiment template.

updateExperimentTemplateResponse_httpStatus :: Lens' UpdateExperimentTemplateResponse Int

The response's http status code.