Amazonka.FMS.GetProtectionStatus

Description

If you created a Shield Advanced policy, returns policy-level attack summary information in the event of a potential DDoS attack. Other policy types are currently unsupported.

Creating a Request

data GetProtectionStatus

See: newGetProtectionStatus smart constructor.

Constructors

GetProtectionStatus'

Fields

memberAccountId :: Maybe Text

The Amazon Web Services account that is in scope of the policy that you want to get the details for.

startTime :: Maybe POSIX

The start of the time period to query for the attacks. This is a timestamp type. The request syntax listing indicates a number type because the default used by Firewall Manager is Unix time in seconds. However, any valid timestamp format is allowed.

nextToken :: Maybe Text

If you specify a value for MaxResults and you have more objects than the number that you specify for MaxResults, Firewall Manager returns a NextToken value in the response, which you can use to retrieve another group of objects. For the second and subsequent GetProtectionStatus requests, specify the value of NextToken from the previous response to get information about another batch of objects.

endTime :: Maybe POSIX

The end of the time period to query for the attacks. This is a timestamp type. The request syntax listing indicates a number type because the default used by Firewall Manager is Unix time in seconds. However, any valid timestamp format is allowed.

maxResults :: Maybe Natural

Specifies the number of objects that you want Firewall Manager to return for this request. If you have more objects than the number that you specify for MaxResults, the response includes a NextToken value that you can use to get another batch of objects.

policyId :: Text

The ID of the policy for which you want to get the attack information.

newGetProtectionStatus

Arguments

:: Text
$sel:policyId:GetProtectionStatus'

-> GetProtectionStatus

Create a value of GetProtectionStatus with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:memberAccountId:GetProtectionStatus', getProtectionStatus_memberAccountId - The Amazon Web Services account that is in scope of the policy that you want to get the details for.

$sel:startTime:GetProtectionStatus', getProtectionStatus_startTime - The start of the time period to query for the attacks. This is a timestamp type. The request syntax listing indicates a number type because the default used by Firewall Manager is Unix time in seconds. However, any valid timestamp format is allowed.

$sel:nextToken:GetProtectionStatus', getProtectionStatus_nextToken - If you specify a value for MaxResults and you have more objects than the number that you specify for MaxResults, Firewall Manager returns a NextToken value in the response, which you can use to retrieve another group of objects. For the second and subsequent GetProtectionStatus requests, specify the value of NextToken from the previous response to get information about another batch of objects.

$sel:endTime:GetProtectionStatus', getProtectionStatus_endTime - The end of the time period to query for the attacks. This is a timestamp type. The request syntax listing indicates a number type because the default used by Firewall Manager is Unix time in seconds. However, any valid timestamp format is allowed.

$sel:maxResults:GetProtectionStatus', getProtectionStatus_maxResults - Specifies the number of objects that you want Firewall Manager to return for this request. If you have more objects than the number that you specify for MaxResults, the response includes a NextToken value that you can use to get another batch of objects.

$sel:policyId:GetProtectionStatus', getProtectionStatus_policyId - The ID of the policy for which you want to get the attack information.

Request Lenses

getProtectionStatus_memberAccountId :: Lens' GetProtectionStatus (Maybe Text)

The Amazon Web Services account that is in scope of the policy that you want to get the details for.

getProtectionStatus_startTime :: Lens' GetProtectionStatus (Maybe UTCTime)

The start of the time period to query for the attacks. This is a timestamp type. The request syntax listing indicates a number type because the default used by Firewall Manager is Unix time in seconds. However, any valid timestamp format is allowed.

getProtectionStatus_nextToken :: Lens' GetProtectionStatus (Maybe Text)

If you specify a value for MaxResults and you have more objects than the number that you specify for MaxResults, Firewall Manager returns a NextToken value in the response, which you can use to retrieve another group of objects. For the second and subsequent GetProtectionStatus requests, specify the value of NextToken from the previous response to get information about another batch of objects.

getProtectionStatus_endTime :: Lens' GetProtectionStatus (Maybe UTCTime)

The end of the time period to query for the attacks. This is a timestamp type. The request syntax listing indicates a number type because the default used by Firewall Manager is Unix time in seconds. However, any valid timestamp format is allowed.

getProtectionStatus_maxResults :: Lens' GetProtectionStatus (Maybe Natural)

Specifies the number of objects that you want Firewall Manager to return for this request. If you have more objects than the number that you specify for MaxResults, the response includes a NextToken value that you can use to get another batch of objects.

getProtectionStatus_policyId :: Lens' GetProtectionStatus Text

The ID of the policy for which you want to get the attack information.

Destructuring the Response

data GetProtectionStatusResponse

See: newGetProtectionStatusResponse smart constructor.

Constructors

GetProtectionStatusResponse'

Fields

data' :: Maybe Text

Details about the attack, including the following:

Attack type

Account ID

ARN of the resource attacked

Start time of the attack

End time of the attack (ongoing attacks will not have an end time)

The details are in JSON format.

adminAccountId :: Maybe Text

The ID of the Firewall Manager administrator account for this policy.

nextToken :: Maybe Text

If you have more objects than the number that you specified for MaxResults in the request, the response includes a NextToken value. To list more objects, submit another GetProtectionStatus request, and specify the NextToken value from the response in the NextToken value in the next request.

Amazon Web Services SDKs provide auto-pagination that identify NextToken in a response and make subsequent request calls automatically on your behalf. However, this feature is not supported by GetProtectionStatus. You must submit subsequent requests with NextToken using your own processes.

serviceType :: Maybe SecurityServiceType

The service type that is protected by the policy. Currently, this is always SHIELD_ADVANCED.

httpStatus :: Int

The response's http status code.

newGetProtectionStatusResponse

Arguments

:: Int
$sel:httpStatus:GetProtectionStatusResponse'

-> GetProtectionStatusResponse

Create a value of GetProtectionStatusResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:data':GetProtectionStatusResponse', getProtectionStatusResponse_data - Details about the attack, including the following:

Attack type

Account ID

ARN of the resource attacked

Start time of the attack

End time of the attack (ongoing attacks will not have an end time)

The details are in JSON format.

$sel:adminAccountId:GetProtectionStatusResponse', getProtectionStatusResponse_adminAccountId - The ID of the Firewall Manager administrator account for this policy.

$sel:nextToken:GetProtectionStatus', getProtectionStatusResponse_nextToken - If you have more objects than the number that you specified for MaxResults in the request, the response includes a NextToken value. To list more objects, submit another GetProtectionStatus request, and specify the NextToken value from the response in the NextToken value in the next request.

Amazon Web Services SDKs provide auto-pagination that identify NextToken in a response and make subsequent request calls automatically on your behalf. However, this feature is not supported by GetProtectionStatus. You must submit subsequent requests with NextToken using your own processes.

$sel:serviceType:GetProtectionStatusResponse', getProtectionStatusResponse_serviceType - The service type that is protected by the policy. Currently, this is always SHIELD_ADVANCED.

$sel:httpStatus:GetProtectionStatusResponse', getProtectionStatusResponse_httpStatus - The response's http status code.

Response Lenses

getProtectionStatusResponse_data :: Lens' GetProtectionStatusResponse (Maybe Text)

Details about the attack, including the following:

Attack type

Account ID

ARN of the resource attacked

Start time of the attack

End time of the attack (ongoing attacks will not have an end time)

The details are in JSON format.

getProtectionStatusResponse_adminAccountId :: Lens' GetProtectionStatusResponse (Maybe Text)

The ID of the Firewall Manager administrator account for this policy.

getProtectionStatusResponse_nextToken :: Lens' GetProtectionStatusResponse (Maybe Text)

If you have more objects than the number that you specified for MaxResults in the request, the response includes a NextToken value. To list more objects, submit another GetProtectionStatus request, and specify the NextToken value from the response in the NextToken value in the next request.

Amazon Web Services SDKs provide auto-pagination that identify NextToken in a response and make subsequent request calls automatically on your behalf. However, this feature is not supported by GetProtectionStatus. You must submit subsequent requests with NextToken using your own processes.

getProtectionStatusResponse_serviceType :: Lens' GetProtectionStatusResponse (Maybe SecurityServiceType)

The service type that is protected by the policy. Currently, this is always SHIELD_ADVANCED.

getProtectionStatusResponse_httpStatus :: Lens' GetProtectionStatusResponse Int

The response's http status code.