Amazonka.FMS.ListProtocolsLists

Description

Returns an array of ProtocolsListDataSummary objects.

This operation returns paginated results.

Creating a Request

data ListProtocolsLists

See: newListProtocolsLists smart constructor.

Constructors

ListProtocolsLists'

Fields

defaultLists :: Maybe Bool

Specifies whether the lists to retrieve are default lists owned by Firewall Manager.

nextToken :: Maybe Text

If you specify a value for MaxResults in your list request, and you have more objects than the maximum, Firewall Manager returns this token in the response. For all but the first request, you provide the token returned by the prior request in the request parameters, to retrieve the next batch of objects.

maxResults :: Natural

The maximum number of objects that you want Firewall Manager to return for this request. If more objects are available, in the response, Firewall Manager provides a NextToken value that you can use in a subsequent call to get the next batch of objects.

If you don't specify this, Firewall Manager returns all available objects.

newListProtocolsLists

Arguments

:: Natural
$sel:maxResults:ListProtocolsLists'

-> ListProtocolsLists

Create a value of ListProtocolsLists with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:defaultLists:ListProtocolsLists', listProtocolsLists_defaultLists - Specifies whether the lists to retrieve are default lists owned by Firewall Manager.

$sel:nextToken:ListProtocolsLists', listProtocolsLists_nextToken - If you specify a value for MaxResults in your list request, and you have more objects than the maximum, Firewall Manager returns this token in the response. For all but the first request, you provide the token returned by the prior request in the request parameters, to retrieve the next batch of objects.

$sel:maxResults:ListProtocolsLists', listProtocolsLists_maxResults - The maximum number of objects that you want Firewall Manager to return for this request. If more objects are available, in the response, Firewall Manager provides a NextToken value that you can use in a subsequent call to get the next batch of objects.

If you don't specify this, Firewall Manager returns all available objects.

Request Lenses

listProtocolsLists_defaultLists :: Lens' ListProtocolsLists (Maybe Bool)

Specifies whether the lists to retrieve are default lists owned by Firewall Manager.

listProtocolsLists_nextToken :: Lens' ListProtocolsLists (Maybe Text)

If you specify a value for MaxResults in your list request, and you have more objects than the maximum, Firewall Manager returns this token in the response. For all but the first request, you provide the token returned by the prior request in the request parameters, to retrieve the next batch of objects.

listProtocolsLists_maxResults :: Lens' ListProtocolsLists Natural

The maximum number of objects that you want Firewall Manager to return for this request. If more objects are available, in the response, Firewall Manager provides a NextToken value that you can use in a subsequent call to get the next batch of objects.

If you don't specify this, Firewall Manager returns all available objects.

Destructuring the Response

data ListProtocolsListsResponse

See: newListProtocolsListsResponse smart constructor.

Constructors

ListProtocolsListsResponse'

Fields

protocolsLists :: Maybe [ProtocolsListDataSummary]

An array of ProtocolsListDataSummary objects.

nextToken :: Maybe Text

If you specify a value for MaxResults in your list request, and you have more objects than the maximum, Firewall Manager returns this token in the response. You can use this token in subsequent requests to retrieve the next batch of objects.

httpStatus :: Int

The response's http status code.

newListProtocolsListsResponse

Arguments

:: Int
$sel:httpStatus:ListProtocolsListsResponse'

-> ListProtocolsListsResponse

Create a value of ListProtocolsListsResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:protocolsLists:ListProtocolsListsResponse', listProtocolsListsResponse_protocolsLists - An array of ProtocolsListDataSummary objects.

$sel:nextToken:ListProtocolsLists', listProtocolsListsResponse_nextToken - If you specify a value for MaxResults in your list request, and you have more objects than the maximum, Firewall Manager returns this token in the response. You can use this token in subsequent requests to retrieve the next batch of objects.

$sel:httpStatus:ListProtocolsListsResponse', listProtocolsListsResponse_httpStatus - The response's http status code.

Response Lenses

listProtocolsListsResponse_protocolsLists :: Lens' ListProtocolsListsResponse (Maybe [ProtocolsListDataSummary])

An array of ProtocolsListDataSummary objects.

listProtocolsListsResponse_nextToken :: Lens' ListProtocolsListsResponse (Maybe Text)

If you specify a value for MaxResults in your list request, and you have more objects than the maximum, Firewall Manager returns this token in the response. You can use this token in subsequent requests to retrieve the next batch of objects.

listProtocolsListsResponse_httpStatus :: Lens' ListProtocolsListsResponse Int

The response's http status code.