Amazonka.FMS.Types.AwsEc2InstanceViolation

Description

Documentation

data AwsEc2InstanceViolation

Violation detail for an EC2 instance resource.

See: newAwsEc2InstanceViolation smart constructor.

Constructors

AwsEc2InstanceViolation'

Fields

violationTarget :: Maybe Text

The resource ID of the EC2 instance.

awsEc2NetworkInterfaceViolations :: Maybe [AwsEc2NetworkInterfaceViolation]

Violation detail for network interfaces associated with the EC2 instance.

newAwsEc2InstanceViolation :: AwsEc2InstanceViolation

Create a value of AwsEc2InstanceViolation with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:violationTarget:AwsEc2InstanceViolation', awsEc2InstanceViolation_violationTarget - The resource ID of the EC2 instance.

$sel:awsEc2NetworkInterfaceViolations:AwsEc2InstanceViolation', awsEc2InstanceViolation_awsEc2NetworkInterfaceViolations - Violation detail for network interfaces associated with the EC2 instance.

awsEc2InstanceViolation_violationTarget :: Lens' AwsEc2InstanceViolation (Maybe Text)

The resource ID of the EC2 instance.

awsEc2InstanceViolation_awsEc2NetworkInterfaceViolations :: Lens' AwsEc2InstanceViolation (Maybe [AwsEc2NetworkInterfaceViolation])

Violation detail for network interfaces associated with the EC2 instance.