Amazonka.FMS.Types.DnsDuplicateRuleGroupViolation

Description

Documentation

data DnsDuplicateRuleGroupViolation

A DNS Firewall rule group that Firewall Manager tried to associate with a VPC is already associated with the VPC and can't be associated again.

See: newDnsDuplicateRuleGroupViolation smart constructor.

Constructors

DnsDuplicateRuleGroupViolation'

Fields

violationTargetDescription :: Maybe Text

A description of the violation that specifies the rule group and VPC.

violationTarget :: Maybe Text

Information about the VPC ID.

newDnsDuplicateRuleGroupViolation :: DnsDuplicateRuleGroupViolation

Create a value of DnsDuplicateRuleGroupViolation with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:violationTargetDescription:DnsDuplicateRuleGroupViolation', dnsDuplicateRuleGroupViolation_violationTargetDescription - A description of the violation that specifies the rule group and VPC.

$sel:violationTarget:DnsDuplicateRuleGroupViolation', dnsDuplicateRuleGroupViolation_violationTarget - Information about the VPC ID.

dnsDuplicateRuleGroupViolation_violationTargetDescription :: Lens' DnsDuplicateRuleGroupViolation (Maybe Text)

A description of the violation that specifies the rule group and VPC.

dnsDuplicateRuleGroupViolation_violationTarget :: Lens' DnsDuplicateRuleGroupViolation (Maybe Text)

Information about the VPC ID.