Amazonka.FMS.Types.NetworkFirewallMissingFirewallViolation

Description

Documentation

data NetworkFirewallMissingFirewallViolation

Violation detail for Network Firewall for a subnet that doesn't have a Firewall Manager managed firewall in its VPC.

See: newNetworkFirewallMissingFirewallViolation smart constructor.

Constructors

NetworkFirewallMissingFirewallViolation'

Fields

targetViolationReason :: Maybe Text

The reason the resource has this violation, if one is available.

availabilityZone :: Maybe Text

The Availability Zone of a violating subnet.

vpc :: Maybe Text

The resource ID of the VPC associated with a violating subnet.

violationTarget :: Maybe Text

The ID of the Network Firewall or VPC resource that's in violation.

newNetworkFirewallMissingFirewallViolation :: NetworkFirewallMissingFirewallViolation

Create a value of NetworkFirewallMissingFirewallViolation with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:targetViolationReason:NetworkFirewallMissingFirewallViolation', networkFirewallMissingFirewallViolation_targetViolationReason - The reason the resource has this violation, if one is available.

$sel:availabilityZone:NetworkFirewallMissingFirewallViolation', networkFirewallMissingFirewallViolation_availabilityZone - The Availability Zone of a violating subnet.

$sel:vpc:NetworkFirewallMissingFirewallViolation', networkFirewallMissingFirewallViolation_vpc - The resource ID of the VPC associated with a violating subnet.

$sel:violationTarget:NetworkFirewallMissingFirewallViolation', networkFirewallMissingFirewallViolation_violationTarget - The ID of the Network Firewall or VPC resource that's in violation.

networkFirewallMissingFirewallViolation_targetViolationReason :: Lens' NetworkFirewallMissingFirewallViolation (Maybe Text)

The reason the resource has this violation, if one is available.

networkFirewallMissingFirewallViolation_availabilityZone :: Lens' NetworkFirewallMissingFirewallViolation (Maybe Text)

The Availability Zone of a violating subnet.

networkFirewallMissingFirewallViolation_vpc :: Lens' NetworkFirewallMissingFirewallViolation (Maybe Text)

The resource ID of the VPC associated with a violating subnet.

networkFirewallMissingFirewallViolation_violationTarget :: Lens' NetworkFirewallMissingFirewallViolation (Maybe Text)

The ID of the Network Firewall or VPC resource that's in violation.