Amazonka.FMS.Types.PolicySummary

Description

Documentation

data PolicySummary

Details of the Firewall Manager policy.

See: newPolicySummary smart constructor.

Constructors

PolicySummary'

Fields

policyName :: Maybe Text

The name of the specified policy.

remediationEnabled :: Maybe Bool

Indicates if the policy should be automatically applied to new resources.

resourceType :: Maybe Text

The type of resource protected by or in scope of the policy. This is in the format shown in the Amazon Web Services Resource Types Reference. For WAF and Shield Advanced, examples include AWS::ElasticLoadBalancingV2::LoadBalancer and AWS::CloudFront::Distribution. For a security group common policy, valid values are AWS::EC2::NetworkInterface and AWS::EC2::Instance. For a security group content audit policy, valid values are AWS::EC2::SecurityGroup, AWS::EC2::NetworkInterface, and AWS::EC2::Instance. For a security group usage audit policy, the value is AWS::EC2::SecurityGroup. For an Network Firewall policy or DNS Firewall policy, the value is AWS::EC2::VPC.

policyId :: Maybe Text

The ID of the specified policy.

deleteUnusedFMManagedResources :: Maybe Bool

Indicates whether Firewall Manager should delete Firewall Manager managed resources, such as web ACLs and security groups, when they are not in use by the Firewall Manager policy. By default, Firewall Manager doesn't delete unused Firewall Manager managed resources. This option is not available for Shield Advanced or WAF Classic policies.

policyArn :: Maybe Text

The Amazon Resource Name (ARN) of the specified policy.

securityServiceType :: Maybe SecurityServiceType

The service that the policy is using to protect the resources. This specifies the type of policy that is created, either an WAF policy, a Shield Advanced policy, or a security group policy.

newPolicySummary :: PolicySummary

Create a value of PolicySummary with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:policyName:PolicySummary', policySummary_policyName - The name of the specified policy.

$sel:remediationEnabled:PolicySummary', policySummary_remediationEnabled - Indicates if the policy should be automatically applied to new resources.

$sel:resourceType:PolicySummary', policySummary_resourceType - The type of resource protected by or in scope of the policy. This is in the format shown in the Amazon Web Services Resource Types Reference. For WAF and Shield Advanced, examples include AWS::ElasticLoadBalancingV2::LoadBalancer and AWS::CloudFront::Distribution. For a security group common policy, valid values are AWS::EC2::NetworkInterface and AWS::EC2::Instance. For a security group content audit policy, valid values are AWS::EC2::SecurityGroup, AWS::EC2::NetworkInterface, and AWS::EC2::Instance. For a security group usage audit policy, the value is AWS::EC2::SecurityGroup. For an Network Firewall policy or DNS Firewall policy, the value is AWS::EC2::VPC.

$sel:policyId:PolicySummary', policySummary_policyId - The ID of the specified policy.

$sel:deleteUnusedFMManagedResources:PolicySummary', policySummary_deleteUnusedFMManagedResources - Indicates whether Firewall Manager should delete Firewall Manager managed resources, such as web ACLs and security groups, when they are not in use by the Firewall Manager policy. By default, Firewall Manager doesn't delete unused Firewall Manager managed resources. This option is not available for Shield Advanced or WAF Classic policies.

$sel:policyArn:PolicySummary', policySummary_policyArn - The Amazon Resource Name (ARN) of the specified policy.

$sel:securityServiceType:PolicySummary', policySummary_securityServiceType - The service that the policy is using to protect the resources. This specifies the type of policy that is created, either an WAF policy, a Shield Advanced policy, or a security group policy.

policySummary_policyName :: Lens' PolicySummary (Maybe Text)

The name of the specified policy.

policySummary_remediationEnabled :: Lens' PolicySummary (Maybe Bool)

Indicates if the policy should be automatically applied to new resources.

policySummary_resourceType :: Lens' PolicySummary (Maybe Text)

The type of resource protected by or in scope of the policy. This is in the format shown in the Amazon Web Services Resource Types Reference. For WAF and Shield Advanced, examples include AWS::ElasticLoadBalancingV2::LoadBalancer and AWS::CloudFront::Distribution. For a security group common policy, valid values are AWS::EC2::NetworkInterface and AWS::EC2::Instance. For a security group content audit policy, valid values are AWS::EC2::SecurityGroup, AWS::EC2::NetworkInterface, and AWS::EC2::Instance. For a security group usage audit policy, the value is AWS::EC2::SecurityGroup. For an Network Firewall policy or DNS Firewall policy, the value is AWS::EC2::VPC.

policySummary_policyId :: Lens' PolicySummary (Maybe Text)

The ID of the specified policy.

policySummary_deleteUnusedFMManagedResources :: Lens' PolicySummary (Maybe Bool)

Indicates whether Firewall Manager should delete Firewall Manager managed resources, such as web ACLs and security groups, when they are not in use by the Firewall Manager policy. By default, Firewall Manager doesn't delete unused Firewall Manager managed resources. This option is not available for Shield Advanced or WAF Classic policies.

policySummary_policyArn :: Lens' PolicySummary (Maybe Text)

The Amazon Resource Name (ARN) of the specified policy.

policySummary_securityServiceType :: Lens' PolicySummary (Maybe SecurityServiceType)

The service that the policy is using to protect the resources. This specifies the type of policy that is created, either an WAF policy, a Shield Advanced policy, or a security group policy.