Amazonka.FMS.Types.ProtocolsListDataSummary

Description

Documentation

data ProtocolsListDataSummary

Details of the Firewall Manager protocols list.

See: newProtocolsListDataSummary smart constructor.

Constructors

ProtocolsListDataSummary'

Fields

protocolsList :: Maybe [Text]

An array of protocols in the Firewall Manager protocols list.

listArn :: Maybe Text

The Amazon Resource Name (ARN) of the specified protocols list.

listId :: Maybe Text

The ID of the specified protocols list.

listName :: Maybe Text

The name of the specified protocols list.

newProtocolsListDataSummary :: ProtocolsListDataSummary

Create a value of ProtocolsListDataSummary with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:protocolsList:ProtocolsListDataSummary', protocolsListDataSummary_protocolsList - An array of protocols in the Firewall Manager protocols list.

$sel:listArn:ProtocolsListDataSummary', protocolsListDataSummary_listArn - The Amazon Resource Name (ARN) of the specified protocols list.

$sel:listId:ProtocolsListDataSummary', protocolsListDataSummary_listId - The ID of the specified protocols list.

$sel:listName:ProtocolsListDataSummary', protocolsListDataSummary_listName - The name of the specified protocols list.

protocolsListDataSummary_protocolsList :: Lens' ProtocolsListDataSummary (Maybe [Text])

An array of protocols in the Firewall Manager protocols list.

protocolsListDataSummary_listArn :: Lens' ProtocolsListDataSummary (Maybe Text)

The Amazon Resource Name (ARN) of the specified protocols list.

protocolsListDataSummary_listId :: Lens' ProtocolsListDataSummary (Maybe Text)

The ID of the specified protocols list.

protocolsListDataSummary_listName :: Lens' ProtocolsListDataSummary (Maybe Text)

The name of the specified protocols list.