Amazonka.FinSpace.CreateEnvironment

Description

Create a new FinSpace environment.

Creating a Request

data CreateEnvironment

See: newCreateEnvironment smart constructor.

Constructors

CreateEnvironment'

Fields

federationParameters :: Maybe FederationParameters

Configuration information when authentication mode is FEDERATED.

federationMode :: Maybe FederationMode

Authentication mode for the environment.

FEDERATED - Users access FinSpace through Single Sign On (SSO) via your Identity provider.

LOCAL - Users access FinSpace via email and password managed within the FinSpace environment.

kmsKeyId :: Maybe Text

The KMS key id to encrypt your data in the FinSpace environment.

description :: Maybe Text

The description of the FinSpace environment to be created.

tags :: Maybe (HashMap Text Text)

Add tags to your FinSpace environment.

name :: Text

The name of the FinSpace environment to be created.

newCreateEnvironment

Arguments

:: Text
$sel:name:CreateEnvironment'

-> CreateEnvironment

Create a value of CreateEnvironment with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:federationParameters:CreateEnvironment', createEnvironment_federationParameters - Configuration information when authentication mode is FEDERATED.

$sel:federationMode:CreateEnvironment', createEnvironment_federationMode - Authentication mode for the environment.

FEDERATED - Users access FinSpace through Single Sign On (SSO) via your Identity provider.

LOCAL - Users access FinSpace via email and password managed within the FinSpace environment.

$sel:kmsKeyId:CreateEnvironment', createEnvironment_kmsKeyId - The KMS key id to encrypt your data in the FinSpace environment.

$sel:description:CreateEnvironment', createEnvironment_description - The description of the FinSpace environment to be created.

$sel:tags:CreateEnvironment', createEnvironment_tags - Add tags to your FinSpace environment.

$sel:name:CreateEnvironment', createEnvironment_name - The name of the FinSpace environment to be created.

Request Lenses

createEnvironment_federationParameters :: Lens' CreateEnvironment (Maybe FederationParameters)

Configuration information when authentication mode is FEDERATED.

createEnvironment_federationMode :: Lens' CreateEnvironment (Maybe FederationMode)

Authentication mode for the environment.

FEDERATED - Users access FinSpace through Single Sign On (SSO) via your Identity provider.

LOCAL - Users access FinSpace via email and password managed within the FinSpace environment.

createEnvironment_kmsKeyId :: Lens' CreateEnvironment (Maybe Text)

The KMS key id to encrypt your data in the FinSpace environment.

createEnvironment_description :: Lens' CreateEnvironment (Maybe Text)

The description of the FinSpace environment to be created.

createEnvironment_tags :: Lens' CreateEnvironment (Maybe (HashMap Text Text))

Add tags to your FinSpace environment.

createEnvironment_name :: Lens' CreateEnvironment Text

The name of the FinSpace environment to be created.

Destructuring the Response

data CreateEnvironmentResponse

See: newCreateEnvironmentResponse smart constructor.

Constructors

CreateEnvironmentResponse'

Fields

environmentUrl :: Maybe Text

The sign-in url for the web application of the FinSpace environment you created.

environmentId :: Maybe Text

The unique identifier for FinSpace environment that you created.

environmentArn :: Maybe Text

The Amazon Resource Name (ARN) of the FinSpace environment that you created.

httpStatus :: Int

The response's http status code.

newCreateEnvironmentResponse

Arguments

:: Int
$sel:httpStatus:CreateEnvironmentResponse'

-> CreateEnvironmentResponse

Create a value of CreateEnvironmentResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:environmentUrl:CreateEnvironmentResponse', createEnvironmentResponse_environmentUrl - The sign-in url for the web application of the FinSpace environment you created.

$sel:environmentId:CreateEnvironmentResponse', createEnvironmentResponse_environmentId - The unique identifier for FinSpace environment that you created.

$sel:environmentArn:CreateEnvironmentResponse', createEnvironmentResponse_environmentArn - The Amazon Resource Name (ARN) of the FinSpace environment that you created.

$sel:httpStatus:CreateEnvironmentResponse', createEnvironmentResponse_httpStatus - The response's http status code.

Response Lenses

createEnvironmentResponse_environmentUrl :: Lens' CreateEnvironmentResponse (Maybe Text)

The sign-in url for the web application of the FinSpace environment you created.

createEnvironmentResponse_environmentId :: Lens' CreateEnvironmentResponse (Maybe Text)

The unique identifier for FinSpace environment that you created.

createEnvironmentResponse_environmentArn :: Lens' CreateEnvironmentResponse (Maybe Text)

The Amazon Resource Name (ARN) of the FinSpace environment that you created.

createEnvironmentResponse_httpStatus :: Lens' CreateEnvironmentResponse Int

The response's http status code.