Amazonka.FinSpace.Lens

Description

Operations

ListEnvironments

listEnvironments_nextToken :: Lens' ListEnvironments (Maybe Text)

A token generated by FinSpace that specifies where to continue pagination if a previous request was truncated. To get the next set of pages, pass in the nextToken value from the response object of the previous page call.

listEnvironments_maxResults :: Lens' ListEnvironments (Maybe Natural)

The maximum number of results to return in this request.

listEnvironmentsResponse_nextToken :: Lens' ListEnvironmentsResponse (Maybe Text)

A token that you can use in a subsequent call to retrieve the next set of results.

listEnvironmentsResponse_environments :: Lens' ListEnvironmentsResponse (Maybe [Environment])

A list of all of your FinSpace environments.

listEnvironmentsResponse_httpStatus :: Lens' ListEnvironmentsResponse Int

The response's http status code.

UpdateEnvironment

updateEnvironment_federationParameters :: Lens' UpdateEnvironment (Maybe FederationParameters)

Undocumented member.

updateEnvironment_federationMode :: Lens' UpdateEnvironment (Maybe FederationMode)

Authentication mode for the environment.

FEDERATED - Users access FinSpace through Single Sign On (SSO) via your Identity provider.

LOCAL - Users access FinSpace via email and password managed within the FinSpace environment.

updateEnvironment_name :: Lens' UpdateEnvironment (Maybe Text)

The name of the environment.

updateEnvironment_description :: Lens' UpdateEnvironment (Maybe Text)

The description of the environment.

updateEnvironment_environmentId :: Lens' UpdateEnvironment Text

The identifier of the FinSpace environment.

updateEnvironmentResponse_environment :: Lens' UpdateEnvironmentResponse (Maybe Environment)

Returns the FinSpace environment object.

updateEnvironmentResponse_httpStatus :: Lens' UpdateEnvironmentResponse Int

The response's http status code.

DeleteEnvironment

deleteEnvironment_environmentId :: Lens' DeleteEnvironment Text

The identifier for the FinSpace environment.

deleteEnvironmentResponse_httpStatus :: Lens' DeleteEnvironmentResponse Int

The response's http status code.

ListTagsForResource

listTagsForResource_resourceArn :: Lens' ListTagsForResource Text

The Amazon Resource Name of the resource.

listTagsForResourceResponse_tags :: Lens' ListTagsForResourceResponse (Maybe (HashMap Text Text))

A list of all tags for a resource.

listTagsForResourceResponse_httpStatus :: Lens' ListTagsForResourceResponse Int

The response's http status code.

TagResource

tagResource_resourceArn :: Lens' TagResource Text

The Amazon Resource Name (ARN) for the resource.

tagResource_tags :: Lens' TagResource (HashMap Text Text)

One or more tags to be assigned to the resource.

tagResourceResponse_httpStatus :: Lens' TagResourceResponse Int

The response's http status code.

GetEnvironment

getEnvironment_environmentId :: Lens' GetEnvironment Text

The identifier of the FinSpace environment.

getEnvironmentResponse_environment :: Lens' GetEnvironmentResponse (Maybe Environment)

The name of the FinSpace environment.

getEnvironmentResponse_httpStatus :: Lens' GetEnvironmentResponse Int

The response's http status code.

UntagResource

untagResource_resourceArn :: Lens' UntagResource Text

A FinSpace resource from which you want to remove a tag or tags. The value for this parameter is an Amazon Resource Name (ARN).

untagResource_tagKeys :: Lens' UntagResource (NonEmpty Text)

The tag keys (names) of one or more tags to be removed.

untagResourceResponse_httpStatus :: Lens' UntagResourceResponse Int

The response's http status code.

CreateEnvironment

createEnvironment_federationParameters :: Lens' CreateEnvironment (Maybe FederationParameters)

Configuration information when authentication mode is FEDERATED.

createEnvironment_federationMode :: Lens' CreateEnvironment (Maybe FederationMode)

Authentication mode for the environment.

FEDERATED - Users access FinSpace through Single Sign On (SSO) via your Identity provider.

LOCAL - Users access FinSpace via email and password managed within the FinSpace environment.

createEnvironment_kmsKeyId :: Lens' CreateEnvironment (Maybe Text)

The KMS key id to encrypt your data in the FinSpace environment.

createEnvironment_description :: Lens' CreateEnvironment (Maybe Text)

The description of the FinSpace environment to be created.

createEnvironment_tags :: Lens' CreateEnvironment (Maybe (HashMap Text Text))

Add tags to your FinSpace environment.

createEnvironment_name :: Lens' CreateEnvironment Text

The name of the FinSpace environment to be created.

createEnvironmentResponse_environmentUrl :: Lens' CreateEnvironmentResponse (Maybe Text)

The sign-in url for the web application of the FinSpace environment you created.

createEnvironmentResponse_environmentId :: Lens' CreateEnvironmentResponse (Maybe Text)

The unique identifier for FinSpace environment that you created.

createEnvironmentResponse_environmentArn :: Lens' CreateEnvironmentResponse (Maybe Text)

The Amazon Resource Name (ARN) of the FinSpace environment that you created.

createEnvironmentResponse_httpStatus :: Lens' CreateEnvironmentResponse Int

The response's http status code.

Types

Environment

environment_status :: Lens' Environment (Maybe EnvironmentStatus)

The current status of creation of the FinSpace environment.

environment_federationParameters :: Lens' Environment (Maybe FederationParameters)

Configuration information when authentication mode is FEDERATED.

environment_dedicatedServiceAccountId :: Lens' Environment (Maybe Text)

The AWS account ID of the dedicated service account associated with your FinSpace environment.

environment_environmentUrl :: Lens' Environment (Maybe Text)

The sign-in url for the web application of your FinSpace environment.

environment_federationMode :: Lens' Environment (Maybe FederationMode)

The authentication mode for the environment.

environment_awsAccountId :: Lens' Environment (Maybe Text)

The ID of the AWS account in which the FinSpace environment is created.

environment_name :: Lens' Environment (Maybe Text)

The name of the FinSpace environment.

environment_kmsKeyId :: Lens' Environment (Maybe Text)

The KMS key id used to encrypt in the FinSpace environment.

environment_environmentId :: Lens' Environment (Maybe Text)

The identifier of the FinSpace environment.

environment_environmentArn :: Lens' Environment (Maybe Text)

The Amazon Resource Name (ARN) of your FinSpace environment.

environment_sageMakerStudioDomainUrl :: Lens' Environment (Maybe Text)

The url of the integrated FinSpace notebook environment in your web application.

environment_description :: Lens' Environment (Maybe Text)

The description of the FinSpace environment.

FederationParameters

federationParameters_samlMetadataURL :: Lens' FederationParameters (Maybe Text)

Provide the metadata URL from your SAML 2.0 compliant identity provider (IdP).

federationParameters_applicationCallBackURL :: Lens' FederationParameters (Maybe Text)

The redirect or sign-in URL that should be entered into the SAML 2.0 compliant identity provider configuration (IdP).

federationParameters_federationURN :: Lens' FederationParameters (Maybe Text)

The Uniform Resource Name (URN). Also referred as Service Provider URN or Audience URI or Service Provider Entity ID.

federationParameters_attributeMap :: Lens' FederationParameters (Maybe (HashMap Text Text))

SAML attribute name and value. The name must always be Email and the value should be set to the attribute definition in which user email is set. For example, name would be Email and value http://schemas.xmlsoap.org/ws/2005/05/identity/claims/emailaddress. Please check your SAML 2.0 compliant identity provider (IdP) documentation for details.

federationParameters_federationProviderName :: Lens' FederationParameters (Maybe Text)

Name of the identity provider (IdP).

federationParameters_samlMetadataDocument :: Lens' FederationParameters (Maybe Text)

SAML 2.0 Metadata document from identity provider (IdP).