Amazonka.FinSpace.ListEnvironments

Description

A list of all of your FinSpace environments.

Creating a Request

data ListEnvironments

See: newListEnvironments smart constructor.

Constructors

ListEnvironments'

Fields

nextToken :: Maybe Text

A token generated by FinSpace that specifies where to continue pagination if a previous request was truncated. To get the next set of pages, pass in the nextToken value from the response object of the previous page call.

maxResults :: Maybe Natural

The maximum number of results to return in this request.

newListEnvironments :: ListEnvironments

Create a value of ListEnvironments with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:nextToken:ListEnvironments', listEnvironments_nextToken - A token generated by FinSpace that specifies where to continue pagination if a previous request was truncated. To get the next set of pages, pass in the nextToken value from the response object of the previous page call.

$sel:maxResults:ListEnvironments', listEnvironments_maxResults - The maximum number of results to return in this request.

Request Lenses

listEnvironments_nextToken :: Lens' ListEnvironments (Maybe Text)

A token generated by FinSpace that specifies where to continue pagination if a previous request was truncated. To get the next set of pages, pass in the nextToken value from the response object of the previous page call.

listEnvironments_maxResults :: Lens' ListEnvironments (Maybe Natural)

The maximum number of results to return in this request.

Destructuring the Response

data ListEnvironmentsResponse

See: newListEnvironmentsResponse smart constructor.

Constructors

ListEnvironmentsResponse'

Fields

nextToken :: Maybe Text

A token that you can use in a subsequent call to retrieve the next set of results.

environments :: Maybe [Environment]

A list of all of your FinSpace environments.

httpStatus :: Int

The response's http status code.

newListEnvironmentsResponse

Arguments

:: Int
$sel:httpStatus:ListEnvironmentsResponse'

-> ListEnvironmentsResponse

Create a value of ListEnvironmentsResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:nextToken:ListEnvironments', listEnvironmentsResponse_nextToken - A token that you can use in a subsequent call to retrieve the next set of results.

$sel:environments:ListEnvironmentsResponse', listEnvironmentsResponse_environments - A list of all of your FinSpace environments.

$sel:httpStatus:ListEnvironmentsResponse', listEnvironmentsResponse_httpStatus - The response's http status code.

Response Lenses

listEnvironmentsResponse_nextToken :: Lens' ListEnvironmentsResponse (Maybe Text)

A token that you can use in a subsequent call to retrieve the next set of results.

listEnvironmentsResponse_environments :: Lens' ListEnvironmentsResponse (Maybe [Environment])

A list of all of your FinSpace environments.

listEnvironmentsResponse_httpStatus :: Lens' ListEnvironmentsResponse Int

The response's http status code.