Amazonka.FinSpaceData

Description

Derived from API version 2020-07-13 of the AWS service descriptions, licensed under Apache 2.0.

The FinSpace APIs let you take actions inside the FinSpace environment.

Service Configuration

defaultService :: Service

API version 2020-07-13 of the Amazon FinSpace Public API SDK configuration.

Errors

Error matchers are designed for use with the functions provided by Control.Exception.Lens. This allows catching (and rethrowing) service specific errors returned by FinSpaceData.

ValidationException

_ValidationException :: AsError a => Getting (First ServiceError) a ServiceError

The input fails to satisfy the constraints specified by an AWS service.

AccessDeniedException

_AccessDeniedException :: AsError a => Getting (First ServiceError) a ServiceError

You do not have sufficient access to perform this action.

ThrottlingException

_ThrottlingException :: AsError a => Getting (First ServiceError) a ServiceError

The request was denied due to request throttling.

InternalServerException

_InternalServerException :: AsError a => Getting (First ServiceError) a ServiceError

The request processing has failed because of an unknown error, exception or failure.

ResourceNotFoundException

_ResourceNotFoundException :: AsError a => Getting (First ServiceError) a ServiceError

One or more resources can't be found.

Waiters

Waiters poll by repeatedly sending a request until some remote success condition configured by the Wait specification is fulfilled. The Wait specification determines how many attempts should be made, in addition to delay and retry strategies.

Operations

Some AWS operations return results that are incomplete and require subsequent requests in order to obtain the entire result set. The process of sending subsequent requests to continue where a previous request left off is called pagination. For example, the ListObjects operation of Amazon S3 returns up to 1000 objects at a time, and you must send subsequent requests with the appropriate Marker in order to retrieve the next page of results.

Operations that have an AWSPager instance can transparently perform subsequent requests, correctly setting Markers and other request facets to iterate through the entire result set of a truncated API operation. Operations which support this have an additional note in the documentation.

Many operations have the ability to filter results on the server side. See the individual operation parameters for details.

CreateChangeset

data CreateChangeset

See: newCreateChangeset smart constructor.

Constructors

CreateChangeset' (Maybe (HashMap Text Text)) (Maybe FormatType) (Maybe (HashMap Text Text)) Text ChangeType SourceType (HashMap Text Text)

newCreateChangeset

Arguments

:: Text
$sel:datasetId:CreateChangeset'

-> ChangeType
$sel:changeType:CreateChangeset'

-> SourceType
$sel:sourceType:CreateChangeset'

-> CreateChangeset

Create a value of CreateChangeset with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:formatParams:CreateChangeset', createChangeset_formatParams - Options that define the structure of the source file(s).

$sel:formatType:CreateChangeset', createChangeset_formatType - Format type of the input files being loaded into the changeset.

$sel:tags:CreateChangeset', createChangeset_tags - Metadata tags to apply to this changeset.

$sel:datasetId:CreateChangeset', createChangeset_datasetId - The unique identifier for the FinSpace dataset in which the changeset will be created.

$sel:changeType:CreateChangeset', createChangeset_changeType - Option to indicate how a changeset will be applied to a dataset.

REPLACE - Changeset will be considered as a replacement to all prior loaded changesets.

APPEND - Changeset will be considered as an addition to the end of all prior loaded changesets.

$sel:sourceType:CreateChangeset', createChangeset_sourceType - Type of the data source from which the files to create the changeset will be sourced.

S3 - Amazon S3.

$sel:sourceParams:CreateChangeset', createChangeset_sourceParams - Source path from which the files to create the changeset will be sourced.

data CreateChangesetResponse

See: newCreateChangesetResponse smart constructor.

Constructors

CreateChangesetResponse' (Maybe ChangesetInfo) Int

newCreateChangesetResponse

Arguments

:: Int
$sel:httpStatus:CreateChangesetResponse'

-> CreateChangesetResponse

Create a value of CreateChangesetResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:changeset:CreateChangesetResponse', createChangesetResponse_changeset - Returns the changeset details.

$sel:httpStatus:CreateChangesetResponse', createChangesetResponse_httpStatus - The response's http status code.

GetProgrammaticAccessCredentials

data GetProgrammaticAccessCredentials

See: newGetProgrammaticAccessCredentials smart constructor.

Constructors

GetProgrammaticAccessCredentials' (Maybe Natural) Text

newGetProgrammaticAccessCredentials

Arguments

:: Text
$sel:environmentId:GetProgrammaticAccessCredentials'

-> GetProgrammaticAccessCredentials

Create a value of GetProgrammaticAccessCredentials with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:durationInMinutes:GetProgrammaticAccessCredentials', getProgrammaticAccessCredentials_durationInMinutes - The time duration in which the credentials remain valid.

$sel:environmentId:GetProgrammaticAccessCredentials', getProgrammaticAccessCredentials_environmentId - The habanero environment identifier.

data GetProgrammaticAccessCredentialsResponse

See: newGetProgrammaticAccessCredentialsResponse smart constructor.

Constructors

GetProgrammaticAccessCredentialsResponse' (Maybe Credentials) (Maybe Natural) Int

newGetProgrammaticAccessCredentialsResponse

Arguments

:: Int
$sel:httpStatus:GetProgrammaticAccessCredentialsResponse'

-> GetProgrammaticAccessCredentialsResponse

Create a value of GetProgrammaticAccessCredentialsResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:credentials:GetProgrammaticAccessCredentialsResponse', getProgrammaticAccessCredentialsResponse_credentials - Returns the programmatic credentials.

$sel:durationInMinutes:GetProgrammaticAccessCredentials', getProgrammaticAccessCredentialsResponse_durationInMinutes - Returns the duration in which the credentials will remain valid.

$sel:httpStatus:GetProgrammaticAccessCredentialsResponse', getProgrammaticAccessCredentialsResponse_httpStatus - The response's http status code.

GetWorkingLocation

data GetWorkingLocation

See: newGetWorkingLocation smart constructor.

Constructors

GetWorkingLocation' (Maybe LocationType)

newGetWorkingLocation :: GetWorkingLocation

Create a value of GetWorkingLocation with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:locationType:GetWorkingLocation', getWorkingLocation_locationType - Specify the type of the working location.

SAGEMAKER - Use the Amazon S3 location as a temporary location to store data content when working with FinSpace Notebooks that run on SageMaker studio.

INGESTION - Use the Amazon S3 location as a staging location to copy your data content and then use the location with the changeset creation operation.

data GetWorkingLocationResponse

See: newGetWorkingLocationResponse smart constructor.

Constructors

GetWorkingLocationResponse' (Maybe Text) (Maybe Text) (Maybe Text) Int

newGetWorkingLocationResponse

Arguments

:: Int
$sel:httpStatus:GetWorkingLocationResponse'

-> GetWorkingLocationResponse

Create a value of GetWorkingLocationResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:s3Path:GetWorkingLocationResponse', getWorkingLocationResponse_s3Path - Returns the Amazon S3 Path for the working location.

$sel:s3Uri:GetWorkingLocationResponse', getWorkingLocationResponse_s3Uri - Returns the Amazon S3 URI for the working location.

$sel:s3Bucket:GetWorkingLocationResponse', getWorkingLocationResponse_s3Bucket - Returns the Amazon S3 bucket name for the working location.

$sel:httpStatus:GetWorkingLocationResponse', getWorkingLocationResponse_httpStatus - The response's http status code.

Types

ChangeType

newtype ChangeType

Constructors

ChangeType'

Fields

fromChangeType :: Text

Bundled Patterns

pattern ChangeType_APPEND :: ChangeType

pattern ChangeType_MODIFY :: ChangeType

pattern ChangeType_REPLACE :: ChangeType

ChangesetStatus

newtype ChangesetStatus

Constructors

ChangesetStatus'

Fields

fromChangesetStatus :: Text

Bundled Patterns

pattern ChangesetStatus_FAILED :: ChangesetStatus

pattern ChangesetStatus_PENDING :: ChangesetStatus

pattern ChangesetStatus_RUNNING :: ChangesetStatus

pattern ChangesetStatus_STOP_REQUESTED :: ChangesetStatus

pattern ChangesetStatus_SUCCESS :: ChangesetStatus

ErrorCategory

newtype ErrorCategory

Constructors

ErrorCategory'

Fields

fromErrorCategory :: Text

Bundled Patterns

pattern ErrorCategory_A_user_recoverable_error_has_occurred :: ErrorCategory

pattern ErrorCategory_An_internal_error_has_occurred :: ErrorCategory

pattern ErrorCategory_Cancelled :: ErrorCategory

pattern ErrorCategory_Missing_required_permission_to_perform_this_request :: ErrorCategory

pattern ErrorCategory_One_or_more_inputs_to_this_request_were_not_found :: ErrorCategory

pattern ErrorCategory_Service_limits_have_been_exceeded :: ErrorCategory

pattern ErrorCategory_The_inputs_to_this_request_are_invalid :: ErrorCategory

pattern ErrorCategory_The_system_temporarily_lacks_sufficient_resources_to_process_the_request :: ErrorCategory

FormatType

newtype FormatType

Constructors

FormatType'

Fields

fromFormatType :: Text

Bundled Patterns

pattern FormatType_CSV :: FormatType

pattern FormatType_JSON :: FormatType

pattern FormatType_PARQUET :: FormatType

pattern FormatType_XML :: FormatType

LocationType

newtype LocationType

Constructors

LocationType'

Fields

fromLocationType :: Text

Bundled Patterns

pattern LocationType_INGESTION :: LocationType

pattern LocationType_SAGEMAKER :: LocationType

SourceType

newtype SourceType

Constructors

SourceType'

Fields

fromSourceType :: Text

Bundled Patterns

pattern SourceType_S3 :: SourceType

ChangesetInfo

data ChangesetInfo

A changeset is unit of data in a dataset.

See: newChangesetInfo smart constructor.

Constructors

ChangesetInfo' (Maybe ChangesetStatus) (Maybe SourceType) (Maybe (HashMap Text Text)) (Maybe (HashMap Text Text)) (Maybe Text) (Maybe Text) (Maybe (HashMap Text Text)) (Maybe POSIX) (Maybe Text) (Maybe FormatType) (Maybe Text) (Maybe ChangeType) (Maybe ErrorInfo) (Maybe Text)

newChangesetInfo :: ChangesetInfo

Create a value of ChangesetInfo with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:status:ChangesetInfo', changesetInfo_status - The status of changeset creation operation.

$sel:sourceType:ChangesetInfo', changesetInfo_sourceType - Type of the data source from which the files to create the changeset are sourced.

S3 - Amazon S3.

$sel:sourceParams:ChangesetInfo', changesetInfo_sourceParams - Source path from which the files to create the changeset are sourced.

$sel:changesetLabels:ChangesetInfo', changesetInfo_changesetLabels - Tags associated with the changeset.

$sel:updatedByChangesetId:ChangesetInfo', changesetInfo_updatedByChangesetId - Unique identifier of the changeset that is updated a changeset.

$sel:datasetId:ChangesetInfo', changesetInfo_datasetId - The unique identifier for the FinSpace dataset in which the changeset is created.

$sel:formatParams:ChangesetInfo', changesetInfo_formatParams - Structure of the source file(s).

$sel:createTimestamp:ChangesetInfo', changesetInfo_createTimestamp - The timestamp at which the changeset was created in FinSpace.

$sel:id:ChangesetInfo', changesetInfo_id - Unique identifier for a changeset.

$sel:formatType:ChangesetInfo', changesetInfo_formatType - Format type of the input files loaded into the changeset.

$sel:updatesChangesetId:ChangesetInfo', changesetInfo_updatesChangesetId - Unique identifier of the changeset that is updated.

$sel:changeType:ChangesetInfo', changesetInfo_changeType - Change type indicates how a changeset is applied to a dataset.

REPLACE - Changeset is considered as a replacement to all prior loaded changesets.

APPEND - Changeset is considered as an addition to the end of all prior loaded changesets.

MODIFY - Changeset is considered as a replacement to a specific prior ingested changeset.

$sel:errorInfo:ChangesetInfo', changesetInfo_errorInfo - The structure with error messages.

$sel:changesetArn:ChangesetInfo', changesetInfo_changesetArn - The ARN identifier of the changeset.

Credentials

data Credentials

Set short term API credentials.

See: newCredentials smart constructor.

Constructors

Credentials' (Maybe Text) (Maybe Text) (Maybe Text)

newCredentials :: Credentials

Create a value of Credentials with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:secretAccessKey:Credentials', credentials_secretAccessKey - The access key.

$sel:sessionToken:Credentials', credentials_sessionToken - The session token.

$sel:accessKeyId:Credentials', credentials_accessKeyId - The access key identifier.

ErrorInfo

data ErrorInfo

Error message.

See: newErrorInfo smart constructor.

Constructors

ErrorInfo' (Maybe ErrorCategory) (Maybe Text)

newErrorInfo :: ErrorInfo

Create a value of ErrorInfo with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:errorCategory:ErrorInfo', errorInfo_errorCategory - The category of the error.

VALIDATION -The inputs to this request are invalid.

SERVICE_QUOTA_EXCEEDED - Service quotas have been exceeded. Please contact AWS support to increase quotas.

ACCESS_DENIED - Missing required permission to perform this request.

RESOURCE_NOT_FOUND - One or more inputs to this request were not found.

THROTTLING - The system temporarily lacks sufficient resources to process the request.

INTERNAL_SERVICE_EXCEPTION - An internal service error has occurred.

CANCELLED - A user recoverable error has occurred.

$sel:errorMessage:ErrorInfo', errorInfo_errorMessage - The text of the error message.