Amazonka.FinSpaceData.GetWorkingLocation

Description

A temporary Amazon S3 location to copy your files from a source location to stage or use as a scratch space in Habanero notebook.

Creating a Request

data GetWorkingLocation

See: newGetWorkingLocation smart constructor.

Constructors

GetWorkingLocation'

Fields

locationType :: Maybe LocationType

Specify the type of the working location.

SAGEMAKER - Use the Amazon S3 location as a temporary location to store data content when working with FinSpace Notebooks that run on SageMaker studio.

INGESTION - Use the Amazon S3 location as a staging location to copy your data content and then use the location with the changeset creation operation.

newGetWorkingLocation :: GetWorkingLocation

Create a value of GetWorkingLocation with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:locationType:GetWorkingLocation', getWorkingLocation_locationType - Specify the type of the working location.

SAGEMAKER - Use the Amazon S3 location as a temporary location to store data content when working with FinSpace Notebooks that run on SageMaker studio.

INGESTION - Use the Amazon S3 location as a staging location to copy your data content and then use the location with the changeset creation operation.

Request Lenses

getWorkingLocation_locationType :: Lens' GetWorkingLocation (Maybe LocationType)

Specify the type of the working location.

SAGEMAKER - Use the Amazon S3 location as a temporary location to store data content when working with FinSpace Notebooks that run on SageMaker studio.

INGESTION - Use the Amazon S3 location as a staging location to copy your data content and then use the location with the changeset creation operation.

Destructuring the Response

data GetWorkingLocationResponse

See: newGetWorkingLocationResponse smart constructor.

Constructors

GetWorkingLocationResponse'

Fields

s3Path :: Maybe Text

Returns the Amazon S3 Path for the working location.

s3Uri :: Maybe Text

Returns the Amazon S3 URI for the working location.

s3Bucket :: Maybe Text

Returns the Amazon S3 bucket name for the working location.

httpStatus :: Int

The response's http status code.

newGetWorkingLocationResponse

Arguments

:: Int
$sel:httpStatus:GetWorkingLocationResponse'

-> GetWorkingLocationResponse

Create a value of GetWorkingLocationResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:s3Path:GetWorkingLocationResponse', getWorkingLocationResponse_s3Path - Returns the Amazon S3 Path for the working location.

$sel:s3Uri:GetWorkingLocationResponse', getWorkingLocationResponse_s3Uri - Returns the Amazon S3 URI for the working location.

$sel:s3Bucket:GetWorkingLocationResponse', getWorkingLocationResponse_s3Bucket - Returns the Amazon S3 bucket name for the working location.

$sel:httpStatus:GetWorkingLocationResponse', getWorkingLocationResponse_httpStatus - The response's http status code.

Response Lenses

getWorkingLocationResponse_s3Path :: Lens' GetWorkingLocationResponse (Maybe Text)

Returns the Amazon S3 Path for the working location.

getWorkingLocationResponse_s3Uri :: Lens' GetWorkingLocationResponse (Maybe Text)

Returns the Amazon S3 URI for the working location.

getWorkingLocationResponse_s3Bucket :: Lens' GetWorkingLocationResponse (Maybe Text)

Returns the Amazon S3 bucket name for the working location.

getWorkingLocationResponse_httpStatus :: Lens' GetWorkingLocationResponse Int

The response's http status code.