Amazonka.FinSpaceData.Lens

Description

Operations

CreateChangeset

createChangeset_formatParams :: Lens' CreateChangeset (Maybe (HashMap Text Text))

Options that define the structure of the source file(s).

createChangeset_formatType :: Lens' CreateChangeset (Maybe FormatType)

Format type of the input files being loaded into the changeset.

createChangeset_tags :: Lens' CreateChangeset (Maybe (HashMap Text Text))

Metadata tags to apply to this changeset.

createChangeset_datasetId :: Lens' CreateChangeset Text

The unique identifier for the FinSpace dataset in which the changeset will be created.

createChangeset_changeType :: Lens' CreateChangeset ChangeType

Option to indicate how a changeset will be applied to a dataset.

REPLACE - Changeset will be considered as a replacement to all prior loaded changesets.

APPEND - Changeset will be considered as an addition to the end of all prior loaded changesets.

createChangeset_sourceType :: Lens' CreateChangeset SourceType

Type of the data source from which the files to create the changeset will be sourced.

S3 - Amazon S3.

createChangeset_sourceParams :: Lens' CreateChangeset (HashMap Text Text)

Source path from which the files to create the changeset will be sourced.

createChangesetResponse_changeset :: Lens' CreateChangesetResponse (Maybe ChangesetInfo)

Returns the changeset details.

createChangesetResponse_httpStatus :: Lens' CreateChangesetResponse Int

The response's http status code.

GetProgrammaticAccessCredentials

getProgrammaticAccessCredentials_durationInMinutes :: Lens' GetProgrammaticAccessCredentials (Maybe Natural)

The time duration in which the credentials remain valid.

getProgrammaticAccessCredentials_environmentId :: Lens' GetProgrammaticAccessCredentials Text

The habanero environment identifier.

getProgrammaticAccessCredentialsResponse_credentials :: Lens' GetProgrammaticAccessCredentialsResponse (Maybe Credentials)

Returns the programmatic credentials.

getProgrammaticAccessCredentialsResponse_durationInMinutes :: Lens' GetProgrammaticAccessCredentialsResponse (Maybe Natural)

Returns the duration in which the credentials will remain valid.

getProgrammaticAccessCredentialsResponse_httpStatus :: Lens' GetProgrammaticAccessCredentialsResponse Int

The response's http status code.

GetWorkingLocation

getWorkingLocation_locationType :: Lens' GetWorkingLocation (Maybe LocationType)

Specify the type of the working location.

SAGEMAKER - Use the Amazon S3 location as a temporary location to store data content when working with FinSpace Notebooks that run on SageMaker studio.

INGESTION - Use the Amazon S3 location as a staging location to copy your data content and then use the location with the changeset creation operation.

getWorkingLocationResponse_s3Path :: Lens' GetWorkingLocationResponse (Maybe Text)

Returns the Amazon S3 Path for the working location.

getWorkingLocationResponse_s3Uri :: Lens' GetWorkingLocationResponse (Maybe Text)

Returns the Amazon S3 URI for the working location.

getWorkingLocationResponse_s3Bucket :: Lens' GetWorkingLocationResponse (Maybe Text)

Returns the Amazon S3 bucket name for the working location.

getWorkingLocationResponse_httpStatus :: Lens' GetWorkingLocationResponse Int

The response's http status code.

Types

ChangesetInfo

changesetInfo_status :: Lens' ChangesetInfo (Maybe ChangesetStatus)

The status of changeset creation operation.

changesetInfo_sourceType :: Lens' ChangesetInfo (Maybe SourceType)

Type of the data source from which the files to create the changeset are sourced.

S3 - Amazon S3.

changesetInfo_sourceParams :: Lens' ChangesetInfo (Maybe (HashMap Text Text))

Source path from which the files to create the changeset are sourced.

changesetInfo_changesetLabels :: Lens' ChangesetInfo (Maybe (HashMap Text Text))

Tags associated with the changeset.

changesetInfo_updatedByChangesetId :: Lens' ChangesetInfo (Maybe Text)

Unique identifier of the changeset that is updated a changeset.

changesetInfo_datasetId :: Lens' ChangesetInfo (Maybe Text)

The unique identifier for the FinSpace dataset in which the changeset is created.

changesetInfo_formatParams :: Lens' ChangesetInfo (Maybe (HashMap Text Text))

Structure of the source file(s).

changesetInfo_createTimestamp :: Lens' ChangesetInfo (Maybe UTCTime)

The timestamp at which the changeset was created in FinSpace.

changesetInfo_id :: Lens' ChangesetInfo (Maybe Text)

Unique identifier for a changeset.

changesetInfo_formatType :: Lens' ChangesetInfo (Maybe FormatType)

Format type of the input files loaded into the changeset.

changesetInfo_updatesChangesetId :: Lens' ChangesetInfo (Maybe Text)

Unique identifier of the changeset that is updated.

changesetInfo_changeType :: Lens' ChangesetInfo (Maybe ChangeType)

Change type indicates how a changeset is applied to a dataset.

REPLACE - Changeset is considered as a replacement to all prior loaded changesets.

APPEND - Changeset is considered as an addition to the end of all prior loaded changesets.

MODIFY - Changeset is considered as a replacement to a specific prior ingested changeset.

changesetInfo_errorInfo :: Lens' ChangesetInfo (Maybe ErrorInfo)

The structure with error messages.

changesetInfo_changesetArn :: Lens' ChangesetInfo (Maybe Text)

The ARN identifier of the changeset.

Credentials

credentials_secretAccessKey :: Lens' Credentials (Maybe Text)

The access key.

credentials_sessionToken :: Lens' Credentials (Maybe Text)

The session token.

credentials_accessKeyId :: Lens' Credentials (Maybe Text)

The access key identifier.

ErrorInfo

errorInfo_errorCategory :: Lens' ErrorInfo (Maybe ErrorCategory)

The category of the error.

VALIDATION -The inputs to this request are invalid.

SERVICE_QUOTA_EXCEEDED - Service quotas have been exceeded. Please contact AWS support to increase quotas.

ACCESS_DENIED - Missing required permission to perform this request.

RESOURCE_NOT_FOUND - One or more inputs to this request were not found.

THROTTLING - The system temporarily lacks sufficient resources to process the request.

INTERNAL_SERVICE_EXCEPTION - An internal service error has occurred.

CANCELLED - A user recoverable error has occurred.

errorInfo_errorMessage :: Lens' ErrorInfo (Maybe Text)

The text of the error message.