Amazonka.FinSpaceData.Types

Description

Service Configuration

defaultService :: Service

API version 2020-07-13 of the Amazon FinSpace Public API SDK configuration.

Errors

_ValidationException :: AsError a => Getting (First ServiceError) a ServiceError

The input fails to satisfy the constraints specified by an AWS service.

_AccessDeniedException :: AsError a => Getting (First ServiceError) a ServiceError

You do not have sufficient access to perform this action.

_ThrottlingException :: AsError a => Getting (First ServiceError) a ServiceError

The request was denied due to request throttling.

_InternalServerException :: AsError a => Getting (First ServiceError) a ServiceError

The request processing has failed because of an unknown error, exception or failure.

_ResourceNotFoundException :: AsError a => Getting (First ServiceError) a ServiceError

One or more resources can't be found.

ChangeType

newtype ChangeType

Constructors

ChangeType'

Fields

fromChangeType :: Text

Bundled Patterns

pattern ChangeType_APPEND :: ChangeType

pattern ChangeType_MODIFY :: ChangeType

pattern ChangeType_REPLACE :: ChangeType

ChangesetStatus

newtype ChangesetStatus

Constructors

ChangesetStatus'

Fields

fromChangesetStatus :: Text

Bundled Patterns

pattern ChangesetStatus_FAILED :: ChangesetStatus

pattern ChangesetStatus_PENDING :: ChangesetStatus

pattern ChangesetStatus_RUNNING :: ChangesetStatus

pattern ChangesetStatus_STOP_REQUESTED :: ChangesetStatus

pattern ChangesetStatus_SUCCESS :: ChangesetStatus

ErrorCategory

newtype ErrorCategory

Constructors

ErrorCategory'

Fields

fromErrorCategory :: Text

Bundled Patterns

pattern ErrorCategory_A_user_recoverable_error_has_occurred :: ErrorCategory

pattern ErrorCategory_An_internal_error_has_occurred :: ErrorCategory

pattern ErrorCategory_Cancelled :: ErrorCategory

pattern ErrorCategory_Missing_required_permission_to_perform_this_request :: ErrorCategory

pattern ErrorCategory_One_or_more_inputs_to_this_request_were_not_found :: ErrorCategory

pattern ErrorCategory_Service_limits_have_been_exceeded :: ErrorCategory

pattern ErrorCategory_The_inputs_to_this_request_are_invalid :: ErrorCategory

pattern ErrorCategory_The_system_temporarily_lacks_sufficient_resources_to_process_the_request :: ErrorCategory

FormatType

newtype FormatType

Constructors

FormatType'

Fields

fromFormatType :: Text

Bundled Patterns

pattern FormatType_CSV :: FormatType

pattern FormatType_JSON :: FormatType

pattern FormatType_PARQUET :: FormatType

pattern FormatType_XML :: FormatType

LocationType

newtype LocationType

Constructors

LocationType'

Fields

fromLocationType :: Text

Bundled Patterns

pattern LocationType_INGESTION :: LocationType

pattern LocationType_SAGEMAKER :: LocationType

SourceType

newtype SourceType

Constructors

SourceType'

Fields

fromSourceType :: Text

Bundled Patterns

pattern SourceType_S3 :: SourceType

ChangesetInfo

data ChangesetInfo

A changeset is unit of data in a dataset.

See: newChangesetInfo smart constructor.

Constructors

ChangesetInfo'

Fields

status :: Maybe ChangesetStatus

The status of changeset creation operation.

sourceType :: Maybe SourceType

Type of the data source from which the files to create the changeset are sourced.

S3 - Amazon S3.

sourceParams :: Maybe (HashMap Text Text)

Source path from which the files to create the changeset are sourced.

changesetLabels :: Maybe (HashMap Text Text)

Tags associated with the changeset.

updatedByChangesetId :: Maybe Text

Unique identifier of the changeset that is updated a changeset.

datasetId :: Maybe Text

The unique identifier for the FinSpace dataset in which the changeset is created.

formatParams :: Maybe (HashMap Text Text)

Structure of the source file(s).

createTimestamp :: Maybe POSIX

The timestamp at which the changeset was created in FinSpace.

id :: Maybe Text

Unique identifier for a changeset.

formatType :: Maybe FormatType

Format type of the input files loaded into the changeset.

updatesChangesetId :: Maybe Text

Unique identifier of the changeset that is updated.

changeType :: Maybe ChangeType

Change type indicates how a changeset is applied to a dataset.

REPLACE - Changeset is considered as a replacement to all prior loaded changesets.

APPEND - Changeset is considered as an addition to the end of all prior loaded changesets.

MODIFY - Changeset is considered as a replacement to a specific prior ingested changeset.

errorInfo :: Maybe ErrorInfo

The structure with error messages.

changesetArn :: Maybe Text

The ARN identifier of the changeset.

newChangesetInfo :: ChangesetInfo

Create a value of ChangesetInfo with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:status:ChangesetInfo', changesetInfo_status - The status of changeset creation operation.

$sel:sourceType:ChangesetInfo', changesetInfo_sourceType - Type of the data source from which the files to create the changeset are sourced.

S3 - Amazon S3.

$sel:sourceParams:ChangesetInfo', changesetInfo_sourceParams - Source path from which the files to create the changeset are sourced.

$sel:changesetLabels:ChangesetInfo', changesetInfo_changesetLabels - Tags associated with the changeset.

$sel:updatedByChangesetId:ChangesetInfo', changesetInfo_updatedByChangesetId - Unique identifier of the changeset that is updated a changeset.

$sel:datasetId:ChangesetInfo', changesetInfo_datasetId - The unique identifier for the FinSpace dataset in which the changeset is created.

$sel:formatParams:ChangesetInfo', changesetInfo_formatParams - Structure of the source file(s).

$sel:createTimestamp:ChangesetInfo', changesetInfo_createTimestamp - The timestamp at which the changeset was created in FinSpace.

$sel:id:ChangesetInfo', changesetInfo_id - Unique identifier for a changeset.

$sel:formatType:ChangesetInfo', changesetInfo_formatType - Format type of the input files loaded into the changeset.

$sel:updatesChangesetId:ChangesetInfo', changesetInfo_updatesChangesetId - Unique identifier of the changeset that is updated.

$sel:changeType:ChangesetInfo', changesetInfo_changeType - Change type indicates how a changeset is applied to a dataset.

REPLACE - Changeset is considered as a replacement to all prior loaded changesets.

APPEND - Changeset is considered as an addition to the end of all prior loaded changesets.

MODIFY - Changeset is considered as a replacement to a specific prior ingested changeset.

$sel:errorInfo:ChangesetInfo', changesetInfo_errorInfo - The structure with error messages.

$sel:changesetArn:ChangesetInfo', changesetInfo_changesetArn - The ARN identifier of the changeset.

changesetInfo_status :: Lens' ChangesetInfo (Maybe ChangesetStatus)

The status of changeset creation operation.

changesetInfo_sourceType :: Lens' ChangesetInfo (Maybe SourceType)

Type of the data source from which the files to create the changeset are sourced.

S3 - Amazon S3.

changesetInfo_sourceParams :: Lens' ChangesetInfo (Maybe (HashMap Text Text))

Source path from which the files to create the changeset are sourced.

changesetInfo_changesetLabels :: Lens' ChangesetInfo (Maybe (HashMap Text Text))

Tags associated with the changeset.

changesetInfo_updatedByChangesetId :: Lens' ChangesetInfo (Maybe Text)

Unique identifier of the changeset that is updated a changeset.

changesetInfo_datasetId :: Lens' ChangesetInfo (Maybe Text)

The unique identifier for the FinSpace dataset in which the changeset is created.

changesetInfo_formatParams :: Lens' ChangesetInfo (Maybe (HashMap Text Text))

Structure of the source file(s).

changesetInfo_createTimestamp :: Lens' ChangesetInfo (Maybe UTCTime)

The timestamp at which the changeset was created in FinSpace.

changesetInfo_id :: Lens' ChangesetInfo (Maybe Text)

Unique identifier for a changeset.

changesetInfo_formatType :: Lens' ChangesetInfo (Maybe FormatType)

Format type of the input files loaded into the changeset.

changesetInfo_updatesChangesetId :: Lens' ChangesetInfo (Maybe Text)

Unique identifier of the changeset that is updated.

changesetInfo_changeType :: Lens' ChangesetInfo (Maybe ChangeType)

Change type indicates how a changeset is applied to a dataset.

REPLACE - Changeset is considered as a replacement to all prior loaded changesets.

APPEND - Changeset is considered as an addition to the end of all prior loaded changesets.

MODIFY - Changeset is considered as a replacement to a specific prior ingested changeset.

changesetInfo_errorInfo :: Lens' ChangesetInfo (Maybe ErrorInfo)

The structure with error messages.

changesetInfo_changesetArn :: Lens' ChangesetInfo (Maybe Text)

The ARN identifier of the changeset.

Credentials

data Credentials

Set short term API credentials.

See: newCredentials smart constructor.

Constructors

Credentials'

Fields

secretAccessKey :: Maybe Text

The access key.

sessionToken :: Maybe Text

The session token.

accessKeyId :: Maybe Text

The access key identifier.

newCredentials :: Credentials

Create a value of Credentials with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:secretAccessKey:Credentials', credentials_secretAccessKey - The access key.

$sel:sessionToken:Credentials', credentials_sessionToken - The session token.

$sel:accessKeyId:Credentials', credentials_accessKeyId - The access key identifier.

credentials_secretAccessKey :: Lens' Credentials (Maybe Text)

The access key.

credentials_sessionToken :: Lens' Credentials (Maybe Text)

The session token.

credentials_accessKeyId :: Lens' Credentials (Maybe Text)

The access key identifier.

ErrorInfo

data ErrorInfo

Error message.

See: newErrorInfo smart constructor.

Constructors

ErrorInfo'

Fields

errorCategory :: Maybe ErrorCategory

The category of the error.

VALIDATION -The inputs to this request are invalid.

SERVICE_QUOTA_EXCEEDED - Service quotas have been exceeded. Please contact AWS support to increase quotas.

ACCESS_DENIED - Missing required permission to perform this request.

RESOURCE_NOT_FOUND - One or more inputs to this request were not found.

THROTTLING - The system temporarily lacks sufficient resources to process the request.

INTERNAL_SERVICE_EXCEPTION - An internal service error has occurred.

CANCELLED - A user recoverable error has occurred.

errorMessage :: Maybe Text

The text of the error message.

newErrorInfo :: ErrorInfo

Create a value of ErrorInfo with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:errorCategory:ErrorInfo', errorInfo_errorCategory - The category of the error.

VALIDATION -The inputs to this request are invalid.

SERVICE_QUOTA_EXCEEDED - Service quotas have been exceeded. Please contact AWS support to increase quotas.

ACCESS_DENIED - Missing required permission to perform this request.

RESOURCE_NOT_FOUND - One or more inputs to this request were not found.

THROTTLING - The system temporarily lacks sufficient resources to process the request.

INTERNAL_SERVICE_EXCEPTION - An internal service error has occurred.

CANCELLED - A user recoverable error has occurred.

$sel:errorMessage:ErrorInfo', errorInfo_errorMessage - The text of the error message.

errorInfo_errorCategory :: Lens' ErrorInfo (Maybe ErrorCategory)

The category of the error.

VALIDATION -The inputs to this request are invalid.

SERVICE_QUOTA_EXCEEDED - Service quotas have been exceeded. Please contact AWS support to increase quotas.

ACCESS_DENIED - Missing required permission to perform this request.

RESOURCE_NOT_FOUND - One or more inputs to this request were not found.

THROTTLING - The system temporarily lacks sufficient resources to process the request.

INTERNAL_SERVICE_EXCEPTION - An internal service error has occurred.

CANCELLED - A user recoverable error has occurred.

errorInfo_errorMessage :: Lens' ErrorInfo (Maybe Text)

The text of the error message.