Amazonka.FinSpaceData.Types.ChangesetInfo

Description

Documentation

data ChangesetInfo

A changeset is unit of data in a dataset.

See: newChangesetInfo smart constructor.

Constructors

ChangesetInfo'

Fields

status :: Maybe ChangesetStatus

The status of changeset creation operation.

sourceType :: Maybe SourceType

Type of the data source from which the files to create the changeset are sourced.

S3 - Amazon S3.

sourceParams :: Maybe (HashMap Text Text)

Source path from which the files to create the changeset are sourced.

changesetLabels :: Maybe (HashMap Text Text)

Tags associated with the changeset.

updatedByChangesetId :: Maybe Text

Unique identifier of the changeset that is updated a changeset.

datasetId :: Maybe Text

The unique identifier for the FinSpace dataset in which the changeset is created.

formatParams :: Maybe (HashMap Text Text)

Structure of the source file(s).

createTimestamp :: Maybe POSIX

The timestamp at which the changeset was created in FinSpace.

id :: Maybe Text

Unique identifier for a changeset.

formatType :: Maybe FormatType

Format type of the input files loaded into the changeset.

updatesChangesetId :: Maybe Text

Unique identifier of the changeset that is updated.

changeType :: Maybe ChangeType

Change type indicates how a changeset is applied to a dataset.

REPLACE - Changeset is considered as a replacement to all prior loaded changesets.

APPEND - Changeset is considered as an addition to the end of all prior loaded changesets.

MODIFY - Changeset is considered as a replacement to a specific prior ingested changeset.

errorInfo :: Maybe ErrorInfo

The structure with error messages.

changesetArn :: Maybe Text

The ARN identifier of the changeset.

newChangesetInfo :: ChangesetInfo

Create a value of ChangesetInfo with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:status:ChangesetInfo', changesetInfo_status - The status of changeset creation operation.

$sel:sourceType:ChangesetInfo', changesetInfo_sourceType - Type of the data source from which the files to create the changeset are sourced.

S3 - Amazon S3.

$sel:sourceParams:ChangesetInfo', changesetInfo_sourceParams - Source path from which the files to create the changeset are sourced.

$sel:changesetLabels:ChangesetInfo', changesetInfo_changesetLabels - Tags associated with the changeset.

$sel:updatedByChangesetId:ChangesetInfo', changesetInfo_updatedByChangesetId - Unique identifier of the changeset that is updated a changeset.

$sel:datasetId:ChangesetInfo', changesetInfo_datasetId - The unique identifier for the FinSpace dataset in which the changeset is created.

$sel:formatParams:ChangesetInfo', changesetInfo_formatParams - Structure of the source file(s).

$sel:createTimestamp:ChangesetInfo', changesetInfo_createTimestamp - The timestamp at which the changeset was created in FinSpace.

$sel:id:ChangesetInfo', changesetInfo_id - Unique identifier for a changeset.

$sel:formatType:ChangesetInfo', changesetInfo_formatType - Format type of the input files loaded into the changeset.

$sel:updatesChangesetId:ChangesetInfo', changesetInfo_updatesChangesetId - Unique identifier of the changeset that is updated.

$sel:changeType:ChangesetInfo', changesetInfo_changeType - Change type indicates how a changeset is applied to a dataset.

REPLACE - Changeset is considered as a replacement to all prior loaded changesets.

APPEND - Changeset is considered as an addition to the end of all prior loaded changesets.

MODIFY - Changeset is considered as a replacement to a specific prior ingested changeset.

$sel:errorInfo:ChangesetInfo', changesetInfo_errorInfo - The structure with error messages.

$sel:changesetArn:ChangesetInfo', changesetInfo_changesetArn - The ARN identifier of the changeset.

changesetInfo_status :: Lens' ChangesetInfo (Maybe ChangesetStatus)

The status of changeset creation operation.

changesetInfo_sourceType :: Lens' ChangesetInfo (Maybe SourceType)

Type of the data source from which the files to create the changeset are sourced.

S3 - Amazon S3.

changesetInfo_sourceParams :: Lens' ChangesetInfo (Maybe (HashMap Text Text))

Source path from which the files to create the changeset are sourced.

changesetInfo_changesetLabels :: Lens' ChangesetInfo (Maybe (HashMap Text Text))

Tags associated with the changeset.

changesetInfo_updatedByChangesetId :: Lens' ChangesetInfo (Maybe Text)

Unique identifier of the changeset that is updated a changeset.

changesetInfo_datasetId :: Lens' ChangesetInfo (Maybe Text)

The unique identifier for the FinSpace dataset in which the changeset is created.

changesetInfo_formatParams :: Lens' ChangesetInfo (Maybe (HashMap Text Text))

Structure of the source file(s).

changesetInfo_createTimestamp :: Lens' ChangesetInfo (Maybe UTCTime)

The timestamp at which the changeset was created in FinSpace.

changesetInfo_id :: Lens' ChangesetInfo (Maybe Text)

Unique identifier for a changeset.

changesetInfo_formatType :: Lens' ChangesetInfo (Maybe FormatType)

Format type of the input files loaded into the changeset.

changesetInfo_updatesChangesetId :: Lens' ChangesetInfo (Maybe Text)

Unique identifier of the changeset that is updated.

changesetInfo_changeType :: Lens' ChangesetInfo (Maybe ChangeType)

Change type indicates how a changeset is applied to a dataset.

REPLACE - Changeset is considered as a replacement to all prior loaded changesets.

APPEND - Changeset is considered as an addition to the end of all prior loaded changesets.

MODIFY - Changeset is considered as a replacement to a specific prior ingested changeset.

changesetInfo_errorInfo :: Lens' ChangesetInfo (Maybe ErrorInfo)

The structure with error messages.

changesetInfo_changesetArn :: Lens' ChangesetInfo (Maybe Text)

The ARN identifier of the changeset.