Amazonka.Firehose.Types.CopyCommand

Description

Documentation

data CopyCommand

Describes a COPY command for Amazon Redshift.

See: newCopyCommand smart constructor.

Constructors

CopyCommand'

Fields

copyOptions :: Maybe Text

Optional parameters to use with the Amazon Redshift COPY command. For more information, see the "Optional Parameters" section of Amazon Redshift COPY command. Some possible examples that would apply to Kinesis Data Firehose are as follows:

delimiter '\t' lzop; - fields are delimited with "\t" (TAB character) and compressed using lzop.

delimiter '|' - fields are delimited with "|" (this is the default delimiter).

delimiter '|' escape - the delimiter should be escaped.

fixedwidth 'venueid:3,venuename:25,venuecity:12,venuestate:2,venueseats:6' - fields are fixed width in the source, with each width specified after every column in the table.

JSON 's3://mybucket/jsonpaths.txt' - data is in JSON format, and the path specified is the format of the data.

For more examples, see Amazon Redshift COPY command examples.

dataTableColumns :: Maybe Text

A comma-separated list of column names.

dataTableName :: Text

The name of the target table. The table must already exist in the database.

newCopyCommand

Arguments

:: Text
$sel:dataTableName:CopyCommand'

-> CopyCommand

Create a value of CopyCommand with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:copyOptions:CopyCommand', copyCommand_copyOptions - Optional parameters to use with the Amazon Redshift COPY command. For more information, see the "Optional Parameters" section of Amazon Redshift COPY command. Some possible examples that would apply to Kinesis Data Firehose are as follows:

delimiter '\t' lzop; - fields are delimited with "\t" (TAB character) and compressed using lzop.

delimiter '|' - fields are delimited with "|" (this is the default delimiter).

delimiter '|' escape - the delimiter should be escaped.

fixedwidth 'venueid:3,venuename:25,venuecity:12,venuestate:2,venueseats:6' - fields are fixed width in the source, with each width specified after every column in the table.

JSON 's3://mybucket/jsonpaths.txt' - data is in JSON format, and the path specified is the format of the data.

For more examples, see Amazon Redshift COPY command examples.

$sel:dataTableColumns:CopyCommand', copyCommand_dataTableColumns - A comma-separated list of column names.

$sel:dataTableName:CopyCommand', copyCommand_dataTableName - The name of the target table. The table must already exist in the database.

copyCommand_copyOptions :: Lens' CopyCommand (Maybe Text)

Optional parameters to use with the Amazon Redshift COPY command. For more information, see the "Optional Parameters" section of Amazon Redshift COPY command. Some possible examples that would apply to Kinesis Data Firehose are as follows:

delimiter '\t' lzop; - fields are delimited with "\t" (TAB character) and compressed using lzop.

delimiter '|' - fields are delimited with "|" (this is the default delimiter).

delimiter '|' escape - the delimiter should be escaped.

fixedwidth 'venueid:3,venuename:25,venuecity:12,venuestate:2,venueseats:6' - fields are fixed width in the source, with each width specified after every column in the table.

JSON 's3://mybucket/jsonpaths.txt' - data is in JSON format, and the path specified is the format of the data.

For more examples, see Amazon Redshift COPY command examples.

copyCommand_dataTableColumns :: Lens' CopyCommand (Maybe Text)

A comma-separated list of column names.

copyCommand_dataTableName :: Lens' CopyCommand Text

The name of the target table. The table must already exist in the database.