Amazonka.Firehose.Types.DataFormatConversionConfiguration

Description

Documentation

data DataFormatConversionConfiguration

Specifies that you want Kinesis Data Firehose to convert data from the JSON format to the Parquet or ORC format before writing it to Amazon S3. Kinesis Data Firehose uses the serializer and deserializer that you specify, in addition to the column information from the AWS Glue table, to deserialize your input data from JSON and then serialize it to the Parquet or ORC format. For more information, see Kinesis Data Firehose Record Format Conversion.

See: newDataFormatConversionConfiguration smart constructor.

Constructors

DataFormatConversionConfiguration'

Fields

outputFormatConfiguration :: Maybe OutputFormatConfiguration

Specifies the serializer that you want Kinesis Data Firehose to use to convert the format of your data to the Parquet or ORC format. This parameter is required if Enabled is set to true.

enabled :: Maybe Bool

Defaults to true. Set it to false if you want to disable format conversion while preserving the configuration details.

schemaConfiguration :: Maybe SchemaConfiguration

Specifies the AWS Glue Data Catalog table that contains the column information. This parameter is required if Enabled is set to true.

inputFormatConfiguration :: Maybe InputFormatConfiguration

Specifies the deserializer that you want Kinesis Data Firehose to use to convert the format of your data from JSON. This parameter is required if Enabled is set to true.

newDataFormatConversionConfiguration :: DataFormatConversionConfiguration

Create a value of DataFormatConversionConfiguration with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:outputFormatConfiguration:DataFormatConversionConfiguration', dataFormatConversionConfiguration_outputFormatConfiguration - Specifies the serializer that you want Kinesis Data Firehose to use to convert the format of your data to the Parquet or ORC format. This parameter is required if Enabled is set to true.

$sel:enabled:DataFormatConversionConfiguration', dataFormatConversionConfiguration_enabled - Defaults to true. Set it to false if you want to disable format conversion while preserving the configuration details.

$sel:schemaConfiguration:DataFormatConversionConfiguration', dataFormatConversionConfiguration_schemaConfiguration - Specifies the AWS Glue Data Catalog table that contains the column information. This parameter is required if Enabled is set to true.

$sel:inputFormatConfiguration:DataFormatConversionConfiguration', dataFormatConversionConfiguration_inputFormatConfiguration - Specifies the deserializer that you want Kinesis Data Firehose to use to convert the format of your data from JSON. This parameter is required if Enabled is set to true.

dataFormatConversionConfiguration_outputFormatConfiguration :: Lens' DataFormatConversionConfiguration (Maybe OutputFormatConfiguration)

Specifies the serializer that you want Kinesis Data Firehose to use to convert the format of your data to the Parquet or ORC format. This parameter is required if Enabled is set to true.

dataFormatConversionConfiguration_enabled :: Lens' DataFormatConversionConfiguration (Maybe Bool)

Defaults to true. Set it to false if you want to disable format conversion while preserving the configuration details.

dataFormatConversionConfiguration_schemaConfiguration :: Lens' DataFormatConversionConfiguration (Maybe SchemaConfiguration)

Specifies the AWS Glue Data Catalog table that contains the column information. This parameter is required if Enabled is set to true.

dataFormatConversionConfiguration_inputFormatConfiguration :: Lens' DataFormatConversionConfiguration (Maybe InputFormatConfiguration)

Specifies the deserializer that you want Kinesis Data Firehose to use to convert the format of your data from JSON. This parameter is required if Enabled is set to true.