Amazonka.Firehose.Types.HttpEndpointDestinationUpdate

Description

Documentation

data HttpEndpointDestinationUpdate

Updates the specified HTTP endpoint destination.

See: newHttpEndpointDestinationUpdate smart constructor.

Constructors

HttpEndpointDestinationUpdate'

Fields

s3BackupMode :: Maybe HttpEndpointS3BackupMode

Describes the S3 bucket backup options for the data that Kinesis Firehose delivers to the HTTP endpoint destination. You can back up all documents (AllData) or only the documents that Kinesis Data Firehose could not deliver to the specified HTTP endpoint destination (FailedDataOnly).

cloudWatchLoggingOptions :: Maybe CloudWatchLoggingOptions

s3Update :: Maybe S3DestinationUpdate

bufferingHints :: Maybe HttpEndpointBufferingHints

Describes buffering options that can be applied to the data before it is delivered to the HTTPS endpoint destination. Kinesis Data Firehose teats these options as hints, and it might choose to use more optimal values. The SizeInMBs and IntervalInSeconds parameters are optional. However, if specify a value for one of them, you must also provide a value for the other.

retryOptions :: Maybe HttpEndpointRetryOptions

Describes the retry behavior in case Kinesis Data Firehose is unable to deliver data to the specified HTTP endpoint destination, or if it doesn't receive a valid acknowledgment of receipt from the specified HTTP endpoint destination.

endpointConfiguration :: Maybe HttpEndpointConfiguration

Describes the configuration of the HTTP endpoint destination.

processingConfiguration :: Maybe ProcessingConfiguration

requestConfiguration :: Maybe HttpEndpointRequestConfiguration

The configuration of the request sent to the HTTP endpoint specified as the destination.

roleARN :: Maybe Text

Kinesis Data Firehose uses this IAM role for all the permissions that the delivery stream needs.

newHttpEndpointDestinationUpdate :: HttpEndpointDestinationUpdate

Create a value of HttpEndpointDestinationUpdate with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:s3BackupMode:HttpEndpointDestinationUpdate', httpEndpointDestinationUpdate_s3BackupMode - Describes the S3 bucket backup options for the data that Kinesis Firehose delivers to the HTTP endpoint destination. You can back up all documents (AllData) or only the documents that Kinesis Data Firehose could not deliver to the specified HTTP endpoint destination (FailedDataOnly).

$sel:cloudWatchLoggingOptions:HttpEndpointDestinationUpdate', httpEndpointDestinationUpdate_cloudWatchLoggingOptions - Undocumented member.

$sel:s3Update:HttpEndpointDestinationUpdate', httpEndpointDestinationUpdate_s3Update - Undocumented member.

$sel:bufferingHints:HttpEndpointDestinationUpdate', httpEndpointDestinationUpdate_bufferingHints - Describes buffering options that can be applied to the data before it is delivered to the HTTPS endpoint destination. Kinesis Data Firehose teats these options as hints, and it might choose to use more optimal values. The SizeInMBs and IntervalInSeconds parameters are optional. However, if specify a value for one of them, you must also provide a value for the other.

$sel:retryOptions:HttpEndpointDestinationUpdate', httpEndpointDestinationUpdate_retryOptions - Describes the retry behavior in case Kinesis Data Firehose is unable to deliver data to the specified HTTP endpoint destination, or if it doesn't receive a valid acknowledgment of receipt from the specified HTTP endpoint destination.

$sel:endpointConfiguration:HttpEndpointDestinationUpdate', httpEndpointDestinationUpdate_endpointConfiguration - Describes the configuration of the HTTP endpoint destination.

$sel:processingConfiguration:HttpEndpointDestinationUpdate', httpEndpointDestinationUpdate_processingConfiguration - Undocumented member.

$sel:requestConfiguration:HttpEndpointDestinationUpdate', httpEndpointDestinationUpdate_requestConfiguration - The configuration of the request sent to the HTTP endpoint specified as the destination.

$sel:roleARN:HttpEndpointDestinationUpdate', httpEndpointDestinationUpdate_roleARN - Kinesis Data Firehose uses this IAM role for all the permissions that the delivery stream needs.

httpEndpointDestinationUpdate_s3BackupMode :: Lens' HttpEndpointDestinationUpdate (Maybe HttpEndpointS3BackupMode)

Describes the S3 bucket backup options for the data that Kinesis Firehose delivers to the HTTP endpoint destination. You can back up all documents (AllData) or only the documents that Kinesis Data Firehose could not deliver to the specified HTTP endpoint destination (FailedDataOnly).

httpEndpointDestinationUpdate_cloudWatchLoggingOptions :: Lens' HttpEndpointDestinationUpdate (Maybe CloudWatchLoggingOptions)

Undocumented member.

httpEndpointDestinationUpdate_s3Update :: Lens' HttpEndpointDestinationUpdate (Maybe S3DestinationUpdate)

Undocumented member.

httpEndpointDestinationUpdate_bufferingHints :: Lens' HttpEndpointDestinationUpdate (Maybe HttpEndpointBufferingHints)

Describes buffering options that can be applied to the data before it is delivered to the HTTPS endpoint destination. Kinesis Data Firehose teats these options as hints, and it might choose to use more optimal values. The SizeInMBs and IntervalInSeconds parameters are optional. However, if specify a value for one of them, you must also provide a value for the other.

httpEndpointDestinationUpdate_retryOptions :: Lens' HttpEndpointDestinationUpdate (Maybe HttpEndpointRetryOptions)

Describes the retry behavior in case Kinesis Data Firehose is unable to deliver data to the specified HTTP endpoint destination, or if it doesn't receive a valid acknowledgment of receipt from the specified HTTP endpoint destination.

httpEndpointDestinationUpdate_endpointConfiguration :: Lens' HttpEndpointDestinationUpdate (Maybe HttpEndpointConfiguration)

Describes the configuration of the HTTP endpoint destination.

httpEndpointDestinationUpdate_processingConfiguration :: Lens' HttpEndpointDestinationUpdate (Maybe ProcessingConfiguration)

Undocumented member.

httpEndpointDestinationUpdate_requestConfiguration :: Lens' HttpEndpointDestinationUpdate (Maybe HttpEndpointRequestConfiguration)

The configuration of the request sent to the HTTP endpoint specified as the destination.

httpEndpointDestinationUpdate_roleARN :: Lens' HttpEndpointDestinationUpdate (Maybe Text)

Kinesis Data Firehose uses this IAM role for all the permissions that the delivery stream needs.