Amazonka.Firehose.Types.RedshiftRetryOptions

Description

Documentation

data RedshiftRetryOptions

Configures retry behavior in case Kinesis Data Firehose is unable to deliver documents to Amazon Redshift.

See: newRedshiftRetryOptions smart constructor.

Constructors

RedshiftRetryOptions'

Fields

durationInSeconds :: Maybe Natural

The length of time during which Kinesis Data Firehose retries delivery after a failure, starting from the initial request and including the first attempt. The default value is 3600 seconds (60 minutes). Kinesis Data Firehose does not retry if the value of DurationInSeconds is 0 (zero) or if the first delivery attempt takes longer than the current value.

newRedshiftRetryOptions :: RedshiftRetryOptions

Create a value of RedshiftRetryOptions with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:durationInSeconds:RedshiftRetryOptions', redshiftRetryOptions_durationInSeconds - The length of time during which Kinesis Data Firehose retries delivery after a failure, starting from the initial request and including the first attempt. The default value is 3600 seconds (60 minutes). Kinesis Data Firehose does not retry if the value of DurationInSeconds is 0 (zero) or if the first delivery attempt takes longer than the current value.

redshiftRetryOptions_durationInSeconds :: Lens' RedshiftRetryOptions (Maybe Natural)

The length of time during which Kinesis Data Firehose retries delivery after a failure, starting from the initial request and including the first attempt. The default value is 3600 seconds (60 minutes). Kinesis Data Firehose does not retry if the value of DurationInSeconds is 0 (zero) or if the first delivery attempt takes longer than the current value.