Amazonka.Firehose.Types.Serializer

Description

Documentation

data Serializer

The serializer that you want Kinesis Data Firehose to use to convert data to the target format before writing it to Amazon S3. Kinesis Data Firehose supports two types of serializers: the ORC SerDe and the Parquet SerDe.

See: newSerializer smart constructor.

Constructors

Serializer'

Fields

orcSerDe :: Maybe OrcSerDe

A serializer to use for converting data to the ORC format before storing it in Amazon S3. For more information, see Apache ORC.

parquetSerDe :: Maybe ParquetSerDe

A serializer to use for converting data to the Parquet format before storing it in Amazon S3. For more information, see Apache Parquet.

newSerializer :: Serializer

Create a value of Serializer with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:orcSerDe:Serializer', serializer_orcSerDe - A serializer to use for converting data to the ORC format before storing it in Amazon S3. For more information, see Apache ORC.

$sel:parquetSerDe:Serializer', serializer_parquetSerDe - A serializer to use for converting data to the Parquet format before storing it in Amazon S3. For more information, see Apache Parquet.

serializer_orcSerDe :: Lens' Serializer (Maybe OrcSerDe)

A serializer to use for converting data to the ORC format before storing it in Amazon S3. For more information, see Apache ORC.

serializer_parquetSerDe :: Lens' Serializer (Maybe ParquetSerDe)

A serializer to use for converting data to the Parquet format before storing it in Amazon S3. For more information, see Apache Parquet.