Amazonka.Firehose.Types.SplunkDestinationConfiguration

Description

Documentation

data SplunkDestinationConfiguration

Describes the configuration of a destination in Splunk.

See: newSplunkDestinationConfiguration smart constructor.

Constructors

SplunkDestinationConfiguration'

Fields

s3BackupMode :: Maybe SplunkS3BackupMode

Defines how documents should be delivered to Amazon S3. When set to FailedEventsOnly, Kinesis Data Firehose writes any data that could not be indexed to the configured Amazon S3 destination. When set to AllEvents, Kinesis Data Firehose delivers all incoming records to Amazon S3, and also writes failed documents to Amazon S3. The default value is FailedEventsOnly.

You can update this backup mode from FailedEventsOnly to AllEvents. You can't update it from AllEvents to FailedEventsOnly.

cloudWatchLoggingOptions :: Maybe CloudWatchLoggingOptions

The Amazon CloudWatch logging options for your delivery stream.

hECAcknowledgmentTimeoutInSeconds :: Maybe Natural

The amount of time that Kinesis Data Firehose waits to receive an acknowledgment from Splunk after it sends it data. At the end of the timeout period, Kinesis Data Firehose either tries to send the data again or considers it an error, based on your retry settings.

retryOptions :: Maybe SplunkRetryOptions

The retry behavior in case Kinesis Data Firehose is unable to deliver data to Splunk, or if it doesn't receive an acknowledgment of receipt from Splunk.

processingConfiguration :: Maybe ProcessingConfiguration

The data processing configuration.

hECEndpoint :: Text

The HTTP Event Collector (HEC) endpoint to which Kinesis Data Firehose sends your data.

hECEndpointType :: HECEndpointType

This type can be either "Raw" or "Event."

hECToken :: Text

This is a GUID that you obtain from your Splunk cluster when you create a new HEC endpoint.

s3Configuration :: S3DestinationConfiguration

The configuration for the backup Amazon S3 location.

newSplunkDestinationConfiguration

Arguments

:: Text
$sel:hECEndpoint:SplunkDestinationConfiguration'

-> HECEndpointType
$sel:hECEndpointType:SplunkDestinationConfiguration'

-> Text
$sel:hECToken:SplunkDestinationConfiguration'

-> S3DestinationConfiguration
$sel:s3Configuration:SplunkDestinationConfiguration'

-> SplunkDestinationConfiguration

Create a value of SplunkDestinationConfiguration with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:s3BackupMode:SplunkDestinationConfiguration', splunkDestinationConfiguration_s3BackupMode - Defines how documents should be delivered to Amazon S3. When set to FailedEventsOnly, Kinesis Data Firehose writes any data that could not be indexed to the configured Amazon S3 destination. When set to AllEvents, Kinesis Data Firehose delivers all incoming records to Amazon S3, and also writes failed documents to Amazon S3. The default value is FailedEventsOnly.

You can update this backup mode from FailedEventsOnly to AllEvents. You can't update it from AllEvents to FailedEventsOnly.

$sel:cloudWatchLoggingOptions:SplunkDestinationConfiguration', splunkDestinationConfiguration_cloudWatchLoggingOptions - The Amazon CloudWatch logging options for your delivery stream.

$sel:hECAcknowledgmentTimeoutInSeconds:SplunkDestinationConfiguration', splunkDestinationConfiguration_hECAcknowledgmentTimeoutInSeconds - The amount of time that Kinesis Data Firehose waits to receive an acknowledgment from Splunk after it sends it data. At the end of the timeout period, Kinesis Data Firehose either tries to send the data again or considers it an error, based on your retry settings.

$sel:retryOptions:SplunkDestinationConfiguration', splunkDestinationConfiguration_retryOptions - The retry behavior in case Kinesis Data Firehose is unable to deliver data to Splunk, or if it doesn't receive an acknowledgment of receipt from Splunk.

$sel:processingConfiguration:SplunkDestinationConfiguration', splunkDestinationConfiguration_processingConfiguration - The data processing configuration.

$sel:hECEndpoint:SplunkDestinationConfiguration', splunkDestinationConfiguration_hECEndpoint - The HTTP Event Collector (HEC) endpoint to which Kinesis Data Firehose sends your data.

$sel:hECEndpointType:SplunkDestinationConfiguration', splunkDestinationConfiguration_hECEndpointType - This type can be either "Raw" or "Event."

$sel:hECToken:SplunkDestinationConfiguration', splunkDestinationConfiguration_hECToken - This is a GUID that you obtain from your Splunk cluster when you create a new HEC endpoint.

$sel:s3Configuration:SplunkDestinationConfiguration', splunkDestinationConfiguration_s3Configuration - The configuration for the backup Amazon S3 location.

splunkDestinationConfiguration_s3BackupMode :: Lens' SplunkDestinationConfiguration (Maybe SplunkS3BackupMode)

Defines how documents should be delivered to Amazon S3. When set to FailedEventsOnly, Kinesis Data Firehose writes any data that could not be indexed to the configured Amazon S3 destination. When set to AllEvents, Kinesis Data Firehose delivers all incoming records to Amazon S3, and also writes failed documents to Amazon S3. The default value is FailedEventsOnly.

You can update this backup mode from FailedEventsOnly to AllEvents. You can't update it from AllEvents to FailedEventsOnly.

splunkDestinationConfiguration_cloudWatchLoggingOptions :: Lens' SplunkDestinationConfiguration (Maybe CloudWatchLoggingOptions)

The Amazon CloudWatch logging options for your delivery stream.

splunkDestinationConfiguration_hECAcknowledgmentTimeoutInSeconds :: Lens' SplunkDestinationConfiguration (Maybe Natural)

The amount of time that Kinesis Data Firehose waits to receive an acknowledgment from Splunk after it sends it data. At the end of the timeout period, Kinesis Data Firehose either tries to send the data again or considers it an error, based on your retry settings.

splunkDestinationConfiguration_retryOptions :: Lens' SplunkDestinationConfiguration (Maybe SplunkRetryOptions)

The retry behavior in case Kinesis Data Firehose is unable to deliver data to Splunk, or if it doesn't receive an acknowledgment of receipt from Splunk.

splunkDestinationConfiguration_processingConfiguration :: Lens' SplunkDestinationConfiguration (Maybe ProcessingConfiguration)

The data processing configuration.

splunkDestinationConfiguration_hECEndpoint :: Lens' SplunkDestinationConfiguration Text

The HTTP Event Collector (HEC) endpoint to which Kinesis Data Firehose sends your data.

splunkDestinationConfiguration_hECEndpointType :: Lens' SplunkDestinationConfiguration HECEndpointType

This type can be either "Raw" or "Event."

splunkDestinationConfiguration_hECToken :: Lens' SplunkDestinationConfiguration Text

This is a GUID that you obtain from your Splunk cluster when you create a new HEC endpoint.

splunkDestinationConfiguration_s3Configuration :: Lens' SplunkDestinationConfiguration S3DestinationConfiguration

The configuration for the backup Amazon S3 location.