Amazonka.Firehose.UntagDeliveryStream

Description

Removes tags from the specified delivery stream. Removed tags are deleted, and you can't recover them after this operation successfully completes.

If you specify a tag that doesn't exist, the operation ignores it.

This operation has a limit of five transactions per second per account.

Creating a Request

data UntagDeliveryStream

See: newUntagDeliveryStream smart constructor.

Constructors

UntagDeliveryStream'

Fields

deliveryStreamName :: Text

The name of the delivery stream.

tagKeys :: NonEmpty Text

A list of tag keys. Each corresponding tag is removed from the delivery stream.

newUntagDeliveryStream

Arguments

:: Text
$sel:deliveryStreamName:UntagDeliveryStream'

-> NonEmpty Text
$sel:tagKeys:UntagDeliveryStream'

-> UntagDeliveryStream

Create a value of UntagDeliveryStream with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:deliveryStreamName:UntagDeliveryStream', untagDeliveryStream_deliveryStreamName - The name of the delivery stream.

$sel:tagKeys:UntagDeliveryStream', untagDeliveryStream_tagKeys - A list of tag keys. Each corresponding tag is removed from the delivery stream.

Request Lenses

untagDeliveryStream_deliveryStreamName :: Lens' UntagDeliveryStream Text

The name of the delivery stream.

untagDeliveryStream_tagKeys :: Lens' UntagDeliveryStream (NonEmpty Text)

A list of tag keys. Each corresponding tag is removed from the delivery stream.

Destructuring the Response

data UntagDeliveryStreamResponse

See: newUntagDeliveryStreamResponse smart constructor.

Constructors

UntagDeliveryStreamResponse'

Fields

httpStatus :: Int

The response's http status code.

newUntagDeliveryStreamResponse

Arguments

:: Int
$sel:httpStatus:UntagDeliveryStreamResponse'

-> UntagDeliveryStreamResponse

Create a value of UntagDeliveryStreamResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:httpStatus:UntagDeliveryStreamResponse', untagDeliveryStreamResponse_httpStatus - The response's http status code.

Response Lenses

untagDeliveryStreamResponse_httpStatus :: Lens' UntagDeliveryStreamResponse Int

The response's http status code.