Amazonka.Firehose.UpdateDestination

Description

Updates the specified destination of the specified delivery stream.

Use this operation to change the destination type (for example, to replace the Amazon S3 destination with Amazon Redshift) or change the parameters associated with a destination (for example, to change the bucket name of the Amazon S3 destination). The update might not occur immediately. The target delivery stream remains active while the configurations are updated, so data writes to the delivery stream can continue during this process. The updated configurations are usually effective within a few minutes.

Switching between Amazon ES and other services is not supported. For an Amazon ES destination, you can only update to another Amazon ES destination.

If the destination type is the same, Kinesis Data Firehose merges the configuration parameters specified with the destination configuration that already exists on the delivery stream. If any of the parameters are not specified in the call, the existing values are retained. For example, in the Amazon S3 destination, if EncryptionConfiguration is not specified, then the existing EncryptionConfiguration is maintained on the destination.

If the destination type is not the same, for example, changing the destination from Amazon S3 to Amazon Redshift, Kinesis Data Firehose does not merge any parameters. In this case, all parameters must be specified.

Kinesis Data Firehose uses CurrentDeliveryStreamVersionId to avoid race conditions and conflicting merges. This is a required field, and the service updates the configuration only if the existing configuration has a version ID that matches. After the update is applied successfully, the version ID is updated, and can be retrieved using DescribeDeliveryStream. Use the new version ID to set CurrentDeliveryStreamVersionId in the next call.

Creating a Request

data UpdateDestination

See: newUpdateDestination smart constructor.

Constructors

UpdateDestination'

Fields

amazonopensearchserviceDestinationUpdate :: Maybe AmazonopensearchserviceDestinationUpdate

splunkDestinationUpdate :: Maybe SplunkDestinationUpdate

Describes an update for a destination in Splunk.

s3DestinationUpdate :: Maybe S3DestinationUpdate

Deprecated

Describes an update for a destination in Amazon S3.

redshiftDestinationUpdate :: Maybe RedshiftDestinationUpdate

Describes an update for a destination in Amazon Redshift.

elasticsearchDestinationUpdate :: Maybe ElasticsearchDestinationUpdate

Describes an update for a destination in Amazon ES.

extendedS3DestinationUpdate :: Maybe ExtendedS3DestinationUpdate

Describes an update for a destination in Amazon S3.

httpEndpointDestinationUpdate :: Maybe HttpEndpointDestinationUpdate

Describes an update to the specified HTTP endpoint destination.

deliveryStreamName :: Text

The name of the delivery stream.

currentDeliveryStreamVersionId :: Text

Obtain this value from the VersionId result of DeliveryStreamDescription. This value is required, and helps the service perform conditional operations. For example, if there is an interleaving update and this value is null, then the update destination fails. After the update is successful, the VersionId value is updated. The service then performs a merge of the old configuration with the new configuration.

destinationId :: Text

The ID of the destination.

newUpdateDestination

Arguments

:: Text
$sel:deliveryStreamName:UpdateDestination'

-> Text
$sel:currentDeliveryStreamVersionId:UpdateDestination'

-> Text
$sel:destinationId:UpdateDestination'

-> UpdateDestination

Create a value of UpdateDestination with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:amazonopensearchserviceDestinationUpdate:UpdateDestination', updateDestination_amazonopensearchserviceDestinationUpdate - Undocumented member.

$sel:splunkDestinationUpdate:UpdateDestination', updateDestination_splunkDestinationUpdate - Describes an update for a destination in Splunk.

$sel:s3DestinationUpdate:UpdateDestination', updateDestination_s3DestinationUpdate - [Deprecated] Describes an update for a destination in Amazon S3.

$sel:redshiftDestinationUpdate:UpdateDestination', updateDestination_redshiftDestinationUpdate - Describes an update for a destination in Amazon Redshift.

$sel:elasticsearchDestinationUpdate:UpdateDestination', updateDestination_elasticsearchDestinationUpdate - Describes an update for a destination in Amazon ES.

$sel:extendedS3DestinationUpdate:UpdateDestination', updateDestination_extendedS3DestinationUpdate - Describes an update for a destination in Amazon S3.

$sel:httpEndpointDestinationUpdate:UpdateDestination', updateDestination_httpEndpointDestinationUpdate - Describes an update to the specified HTTP endpoint destination.

$sel:deliveryStreamName:UpdateDestination', updateDestination_deliveryStreamName - The name of the delivery stream.

$sel:currentDeliveryStreamVersionId:UpdateDestination', updateDestination_currentDeliveryStreamVersionId - Obtain this value from the VersionId result of DeliveryStreamDescription. This value is required, and helps the service perform conditional operations. For example, if there is an interleaving update and this value is null, then the update destination fails. After the update is successful, the VersionId value is updated. The service then performs a merge of the old configuration with the new configuration.

$sel:destinationId:UpdateDestination', updateDestination_destinationId - The ID of the destination.

Request Lenses

updateDestination_amazonopensearchserviceDestinationUpdate :: Lens' UpdateDestination (Maybe AmazonopensearchserviceDestinationUpdate)

Undocumented member.

updateDestination_splunkDestinationUpdate :: Lens' UpdateDestination (Maybe SplunkDestinationUpdate)

Describes an update for a destination in Splunk.

updateDestination_s3DestinationUpdate :: Lens' UpdateDestination (Maybe S3DestinationUpdate)

Deprecated

Describes an update for a destination in Amazon S3.

updateDestination_redshiftDestinationUpdate :: Lens' UpdateDestination (Maybe RedshiftDestinationUpdate)

Describes an update for a destination in Amazon Redshift.

updateDestination_elasticsearchDestinationUpdate :: Lens' UpdateDestination (Maybe ElasticsearchDestinationUpdate)

Describes an update for a destination in Amazon ES.

updateDestination_extendedS3DestinationUpdate :: Lens' UpdateDestination (Maybe ExtendedS3DestinationUpdate)

Describes an update for a destination in Amazon S3.

updateDestination_httpEndpointDestinationUpdate :: Lens' UpdateDestination (Maybe HttpEndpointDestinationUpdate)

Describes an update to the specified HTTP endpoint destination.

updateDestination_deliveryStreamName :: Lens' UpdateDestination Text

The name of the delivery stream.

updateDestination_currentDeliveryStreamVersionId :: Lens' UpdateDestination Text

Obtain this value from the VersionId result of DeliveryStreamDescription. This value is required, and helps the service perform conditional operations. For example, if there is an interleaving update and this value is null, then the update destination fails. After the update is successful, the VersionId value is updated. The service then performs a merge of the old configuration with the new configuration.

updateDestination_destinationId :: Lens' UpdateDestination Text

The ID of the destination.

Destructuring the Response

data UpdateDestinationResponse

See: newUpdateDestinationResponse smart constructor.

Constructors

UpdateDestinationResponse'

Fields

httpStatus :: Int

The response's http status code.

newUpdateDestinationResponse

Arguments

:: Int
$sel:httpStatus:UpdateDestinationResponse'

-> UpdateDestinationResponse

Create a value of UpdateDestinationResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:httpStatus:UpdateDestinationResponse', updateDestinationResponse_httpStatus - The response's http status code.

Response Lenses

updateDestinationResponse_httpStatus :: Lens' UpdateDestinationResponse Int

The response's http status code.