Amazonka.Forecast.CreateDatasetImportJob

Description

Imports your training data to an Amazon Forecast dataset. You provide the location of your training data in an Amazon Simple Storage Service (Amazon S3) bucket and the Amazon Resource Name (ARN) of the dataset that you want to import the data to.

You must specify a DataSource object that includes an AWS Identity and Access Management (IAM) role that Amazon Forecast can assume to access the data, as Amazon Forecast makes a copy of your data and processes it in an internal AWS system. For more information, see aws-forecast-iam-roles.

The training data must be in CSV format. The delimiter must be a comma (,).

You can specify the path to a specific CSV file, the S3 bucket, or to a folder in the S3 bucket. For the latter two cases, Amazon Forecast imports all files up to the limit of 10,000 files.

Because dataset imports are not aggregated, your most recent dataset import is the one that is used when training a predictor or generating a forecast. Make sure that your most recent dataset import contains all of the data you want to model off of, and not just the new data collected since the previous import.

To get a list of all your dataset import jobs, filtered by specified criteria, use the ListDatasetImportJobs operation.

Creating a Request

data CreateDatasetImportJob

See: newCreateDatasetImportJob smart constructor.

Constructors

CreateDatasetImportJob'

Fields

timestampFormat :: Maybe Text

The format of timestamps in the dataset. The format that you specify depends on the DataFrequency specified when the dataset was created. The following formats are supported

"yyyy-MM-dd"

For the following data frequencies: Y, M, W, and D

"yyyy-MM-dd HH:mm:ss"

For the following data frequencies: H, 30min, 15min, and 1min; and optionally, for: Y, M, W, and D

If the format isn't specified, Amazon Forecast expects the format to be "yyyy-MM-dd HH:mm:ss".

useGeolocationForTimeZone :: Maybe Bool

Automatically derive time zone information from the geolocation attribute. This option is ideal for datasets that contain timestamps in multiple time zones and those timestamps are expressed in local time.

geolocationFormat :: Maybe Text

The format of the geolocation attribute. The geolocation attribute can be formatted in one of two ways:

LAT_LONG - the latitude and longitude in decimal format (Example: 47.61_-122.33).

CC_POSTALCODE (US Only) - the country code (US), followed by the 5-digit ZIP code (Example: US_98121).

timeZone :: Maybe Text

A single time zone for every item in your dataset. This option is ideal for datasets with all timestamps within a single time zone, or if all timestamps are normalized to a single time zone.

Refer to the Joda-Time API for a complete list of valid time zone names.

tags :: Maybe [Tag]

The optional metadata that you apply to the dataset import job to help you categorize and organize them. Each tag consists of a key and an optional value, both of which you define.

The following basic restrictions apply to tags:

Maximum number of tags per resource - 50.

For each resource, each tag key must be unique, and each tag key can have only one value.

Maximum key length - 128 Unicode characters in UTF-8.

Maximum value length - 256 Unicode characters in UTF-8.

If your tagging schema is used across multiple services and resources, remember that other services may have restrictions on allowed characters. Generally allowed characters are: letters, numbers, and spaces representable in UTF-8, and the following characters: + - = . _ : / @.

Tag keys and values are case sensitive.

Do not use aws:, AWS:, or any upper or lowercase combination of such as a prefix for keys as it is reserved for AWS use. You cannot edit or delete tag keys with this prefix. Values can have this prefix. If a tag value has aws as its prefix but the key does not, then Forecast considers it to be a user tag and will count against the limit of 50 tags. Tags with only the key prefix of aws do not count against your tags per resource limit.

datasetImportJobName :: Text

The name for the dataset import job. We recommend including the current timestamp in the name, for example, 20190721DatasetImport. This can help you avoid getting a ResourceAlreadyExistsException exception.

datasetArn :: Text

The Amazon Resource Name (ARN) of the Amazon Forecast dataset that you want to import data to.

dataSource :: DataSource

The location of the training data to import and an AWS Identity and Access Management (IAM) role that Amazon Forecast can assume to access the data. The training data must be stored in an Amazon S3 bucket.

If encryption is used, DataSource must include an AWS Key Management Service (KMS) key and the IAM role must allow Amazon Forecast permission to access the key. The KMS key and IAM role must match those specified in the EncryptionConfig parameter of the CreateDataset operation.

newCreateDatasetImportJob

Arguments

:: Text
$sel:datasetImportJobName:CreateDatasetImportJob'

-> Text
$sel:datasetArn:CreateDatasetImportJob'

-> DataSource
$sel:dataSource:CreateDatasetImportJob'

-> CreateDatasetImportJob

Create a value of CreateDatasetImportJob with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:timestampFormat:CreateDatasetImportJob', createDatasetImportJob_timestampFormat - The format of timestamps in the dataset. The format that you specify depends on the DataFrequency specified when the dataset was created. The following formats are supported

"yyyy-MM-dd"

For the following data frequencies: Y, M, W, and D

"yyyy-MM-dd HH:mm:ss"

For the following data frequencies: H, 30min, 15min, and 1min; and optionally, for: Y, M, W, and D

If the format isn't specified, Amazon Forecast expects the format to be "yyyy-MM-dd HH:mm:ss".

$sel:useGeolocationForTimeZone:CreateDatasetImportJob', createDatasetImportJob_useGeolocationForTimeZone - Automatically derive time zone information from the geolocation attribute. This option is ideal for datasets that contain timestamps in multiple time zones and those timestamps are expressed in local time.

$sel:geolocationFormat:CreateDatasetImportJob', createDatasetImportJob_geolocationFormat - The format of the geolocation attribute. The geolocation attribute can be formatted in one of two ways:

LAT_LONG - the latitude and longitude in decimal format (Example: 47.61_-122.33).

CC_POSTALCODE (US Only) - the country code (US), followed by the 5-digit ZIP code (Example: US_98121).

$sel:timeZone:CreateDatasetImportJob', createDatasetImportJob_timeZone - A single time zone for every item in your dataset. This option is ideal for datasets with all timestamps within a single time zone, or if all timestamps are normalized to a single time zone.

Refer to the Joda-Time API for a complete list of valid time zone names.

$sel:tags:CreateDatasetImportJob', createDatasetImportJob_tags - The optional metadata that you apply to the dataset import job to help you categorize and organize them. Each tag consists of a key and an optional value, both of which you define.

The following basic restrictions apply to tags:

Maximum number of tags per resource - 50.

For each resource, each tag key must be unique, and each tag key can have only one value.

Maximum key length - 128 Unicode characters in UTF-8.

Maximum value length - 256 Unicode characters in UTF-8.

If your tagging schema is used across multiple services and resources, remember that other services may have restrictions on allowed characters. Generally allowed characters are: letters, numbers, and spaces representable in UTF-8, and the following characters: + - = . _ : / @.

Tag keys and values are case sensitive.

Do not use aws:, AWS:, or any upper or lowercase combination of such as a prefix for keys as it is reserved for AWS use. You cannot edit or delete tag keys with this prefix. Values can have this prefix. If a tag value has aws as its prefix but the key does not, then Forecast considers it to be a user tag and will count against the limit of 50 tags. Tags with only the key prefix of aws do not count against your tags per resource limit.

$sel:datasetImportJobName:CreateDatasetImportJob', createDatasetImportJob_datasetImportJobName - The name for the dataset import job. We recommend including the current timestamp in the name, for example, 20190721DatasetImport. This can help you avoid getting a ResourceAlreadyExistsException exception.

$sel:datasetArn:CreateDatasetImportJob', createDatasetImportJob_datasetArn - The Amazon Resource Name (ARN) of the Amazon Forecast dataset that you want to import data to.

$sel:dataSource:CreateDatasetImportJob', createDatasetImportJob_dataSource - The location of the training data to import and an AWS Identity and Access Management (IAM) role that Amazon Forecast can assume to access the data. The training data must be stored in an Amazon S3 bucket.

If encryption is used, DataSource must include an AWS Key Management Service (KMS) key and the IAM role must allow Amazon Forecast permission to access the key. The KMS key and IAM role must match those specified in the EncryptionConfig parameter of the CreateDataset operation.

Request Lenses

createDatasetImportJob_timestampFormat :: Lens' CreateDatasetImportJob (Maybe Text)

The format of timestamps in the dataset. The format that you specify depends on the DataFrequency specified when the dataset was created. The following formats are supported

"yyyy-MM-dd"

For the following data frequencies: Y, M, W, and D

"yyyy-MM-dd HH:mm:ss"

For the following data frequencies: H, 30min, 15min, and 1min; and optionally, for: Y, M, W, and D

If the format isn't specified, Amazon Forecast expects the format to be "yyyy-MM-dd HH:mm:ss".

createDatasetImportJob_useGeolocationForTimeZone :: Lens' CreateDatasetImportJob (Maybe Bool)

Automatically derive time zone information from the geolocation attribute. This option is ideal for datasets that contain timestamps in multiple time zones and those timestamps are expressed in local time.

createDatasetImportJob_geolocationFormat :: Lens' CreateDatasetImportJob (Maybe Text)

The format of the geolocation attribute. The geolocation attribute can be formatted in one of two ways:

LAT_LONG - the latitude and longitude in decimal format (Example: 47.61_-122.33).

CC_POSTALCODE (US Only) - the country code (US), followed by the 5-digit ZIP code (Example: US_98121).

createDatasetImportJob_timeZone :: Lens' CreateDatasetImportJob (Maybe Text)

A single time zone for every item in your dataset. This option is ideal for datasets with all timestamps within a single time zone, or if all timestamps are normalized to a single time zone.

Refer to the Joda-Time API for a complete list of valid time zone names.

createDatasetImportJob_tags :: Lens' CreateDatasetImportJob (Maybe [Tag])

The optional metadata that you apply to the dataset import job to help you categorize and organize them. Each tag consists of a key and an optional value, both of which you define.

The following basic restrictions apply to tags:

Maximum number of tags per resource - 50.

For each resource, each tag key must be unique, and each tag key can have only one value.

Maximum key length - 128 Unicode characters in UTF-8.

Maximum value length - 256 Unicode characters in UTF-8.

If your tagging schema is used across multiple services and resources, remember that other services may have restrictions on allowed characters. Generally allowed characters are: letters, numbers, and spaces representable in UTF-8, and the following characters: + - = . _ : / @.

Tag keys and values are case sensitive.

Do not use aws:, AWS:, or any upper or lowercase combination of such as a prefix for keys as it is reserved for AWS use. You cannot edit or delete tag keys with this prefix. Values can have this prefix. If a tag value has aws as its prefix but the key does not, then Forecast considers it to be a user tag and will count against the limit of 50 tags. Tags with only the key prefix of aws do not count against your tags per resource limit.

createDatasetImportJob_datasetImportJobName :: Lens' CreateDatasetImportJob Text

The name for the dataset import job. We recommend including the current timestamp in the name, for example, 20190721DatasetImport. This can help you avoid getting a ResourceAlreadyExistsException exception.

createDatasetImportJob_datasetArn :: Lens' CreateDatasetImportJob Text

The Amazon Resource Name (ARN) of the Amazon Forecast dataset that you want to import data to.

createDatasetImportJob_dataSource :: Lens' CreateDatasetImportJob DataSource

The location of the training data to import and an AWS Identity and Access Management (IAM) role that Amazon Forecast can assume to access the data. The training data must be stored in an Amazon S3 bucket.

If encryption is used, DataSource must include an AWS Key Management Service (KMS) key and the IAM role must allow Amazon Forecast permission to access the key. The KMS key and IAM role must match those specified in the EncryptionConfig parameter of the CreateDataset operation.

Destructuring the Response

data CreateDatasetImportJobResponse

See: newCreateDatasetImportJobResponse smart constructor.

Constructors

CreateDatasetImportJobResponse'

Fields

datasetImportJobArn :: Maybe Text

The Amazon Resource Name (ARN) of the dataset import job.

httpStatus :: Int

The response's http status code.

newCreateDatasetImportJobResponse

Arguments

:: Int
$sel:httpStatus:CreateDatasetImportJobResponse'

-> CreateDatasetImportJobResponse

Create a value of CreateDatasetImportJobResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:datasetImportJobArn:CreateDatasetImportJobResponse', createDatasetImportJobResponse_datasetImportJobArn - The Amazon Resource Name (ARN) of the dataset import job.

$sel:httpStatus:CreateDatasetImportJobResponse', createDatasetImportJobResponse_httpStatus - The response's http status code.

Response Lenses

createDatasetImportJobResponse_datasetImportJobArn :: Lens' CreateDatasetImportJobResponse (Maybe Text)

The Amazon Resource Name (ARN) of the dataset import job.

createDatasetImportJobResponse_httpStatus :: Lens' CreateDatasetImportJobResponse Int

The response's http status code.