Amazonka.Forecast.DeleteDatasetImportJob

Description

Deletes a dataset import job created using the CreateDatasetImportJob operation. You can delete only dataset import jobs that have a status of ACTIVE or CREATE_FAILED. To get the status, use the DescribeDatasetImportJob operation.

Creating a Request

data DeleteDatasetImportJob

See: newDeleteDatasetImportJob smart constructor.

Constructors

DeleteDatasetImportJob'

Fields

datasetImportJobArn :: Text

The Amazon Resource Name (ARN) of the dataset import job to delete.

newDeleteDatasetImportJob

Arguments

:: Text
$sel:datasetImportJobArn:DeleteDatasetImportJob'

-> DeleteDatasetImportJob

Create a value of DeleteDatasetImportJob with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:datasetImportJobArn:DeleteDatasetImportJob', deleteDatasetImportJob_datasetImportJobArn - The Amazon Resource Name (ARN) of the dataset import job to delete.

Request Lenses

deleteDatasetImportJob_datasetImportJobArn :: Lens' DeleteDatasetImportJob Text

The Amazon Resource Name (ARN) of the dataset import job to delete.

Destructuring the Response

data DeleteDatasetImportJobResponse

See: newDeleteDatasetImportJobResponse smart constructor.

Constructors

DeleteDatasetImportJobResponse'

newDeleteDatasetImportJobResponse :: DeleteDatasetImportJobResponse

Create a value of DeleteDatasetImportJobResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.