Amazonka.Forecast.DeleteResourceTree

Description

Deletes an entire resource tree. This operation will delete the parent resource and its child resources.

Child resources are resources that were created from another resource. For example, when a forecast is generated from a predictor, the forecast is the child resource and the predictor is the parent resource.

Amazon Forecast resources possess the following parent-child resource hierarchies:

Dataset: dataset import jobs

Dataset Group: predictors, predictor backtest export jobs, forecasts, forecast export jobs

Predictor: predictor backtest export jobs, forecasts, forecast export jobs

Forecast: forecast export jobs

DeleteResourceTree will only delete Amazon Forecast resources, and will not delete datasets or exported files stored in Amazon S3.

Creating a Request

data DeleteResourceTree

See: newDeleteResourceTree smart constructor.

Constructors

DeleteResourceTree'

Fields

resourceArn :: Text

The Amazon Resource Name (ARN) of the parent resource to delete. All child resources of the parent resource will also be deleted.

newDeleteResourceTree

Arguments

:: Text
$sel:resourceArn:DeleteResourceTree'

-> DeleteResourceTree

Create a value of DeleteResourceTree with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:resourceArn:DeleteResourceTree', deleteResourceTree_resourceArn - The Amazon Resource Name (ARN) of the parent resource to delete. All child resources of the parent resource will also be deleted.

Request Lenses

deleteResourceTree_resourceArn :: Lens' DeleteResourceTree Text

The Amazon Resource Name (ARN) of the parent resource to delete. All child resources of the parent resource will also be deleted.

Destructuring the Response

data DeleteResourceTreeResponse

See: newDeleteResourceTreeResponse smart constructor.

Constructors

DeleteResourceTreeResponse'

newDeleteResourceTreeResponse :: DeleteResourceTreeResponse

Create a value of DeleteResourceTreeResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.