Amazonka.Forecast.DescribeDatasetImportJob

Description

Describes a dataset import job created using the CreateDatasetImportJob operation.

In addition to listing the parameters provided in the CreateDatasetImportJob request, this operation includes the following properties:

CreationTime

LastModificationTime

DataSize

FieldStatistics

Status

Message - If an error occurred, information about the error.

Creating a Request

data DescribeDatasetImportJob

See: newDescribeDatasetImportJob smart constructor.

Constructors

DescribeDatasetImportJob'

Fields

datasetImportJobArn :: Text

The Amazon Resource Name (ARN) of the dataset import job.

newDescribeDatasetImportJob

Arguments

:: Text
$sel:datasetImportJobArn:DescribeDatasetImportJob'

-> DescribeDatasetImportJob

Create a value of DescribeDatasetImportJob with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:datasetImportJobArn:DescribeDatasetImportJob', describeDatasetImportJob_datasetImportJobArn - The Amazon Resource Name (ARN) of the dataset import job.

Request Lenses

describeDatasetImportJob_datasetImportJobArn :: Lens' DescribeDatasetImportJob Text

The Amazon Resource Name (ARN) of the dataset import job.

Destructuring the Response

data DescribeDatasetImportJobResponse

See: newDescribeDatasetImportJobResponse smart constructor.

Constructors

DescribeDatasetImportJobResponse'

Fields

creationTime :: Maybe POSIX

When the dataset import job was created.

status :: Maybe Text

The status of the dataset import job. States include:

ACTIVE

CREATE_PENDING, CREATE_IN_PROGRESS, CREATE_FAILED

DELETE_PENDING, DELETE_IN_PROGRESS, DELETE_FAILED

CREATE_STOPPING, CREATE_STOPPED

datasetImportJobName :: Maybe Text

The name of the dataset import job.

datasetArn :: Maybe Text

The Amazon Resource Name (ARN) of the dataset that the training data was imported to.

timestampFormat :: Maybe Text

The format of timestamps in the dataset. The format that you specify depends on the DataFrequency specified when the dataset was created. The following formats are supported

"yyyy-MM-dd"

For the following data frequencies: Y, M, W, and D

"yyyy-MM-dd HH:mm:ss"

For the following data frequencies: H, 30min, 15min, and 1min; and optionally, for: Y, M, W, and D

dataSize :: Maybe Double

The size of the dataset in gigabytes (GB) after the import job has finished.

estimatedTimeRemainingInMinutes :: Maybe Integer

The estimated time remaining in minutes for the dataset import job to complete.

fieldStatistics :: Maybe (HashMap Text Statistics)

Statistical information about each field in the input data.

dataSource :: Maybe DataSource

The location of the training data to import and an AWS Identity and Access Management (IAM) role that Amazon Forecast can assume to access the data.

If encryption is used, DataSource includes an AWS Key Management Service (KMS) key.

datasetImportJobArn :: Maybe Text

The ARN of the dataset import job.

useGeolocationForTimeZone :: Maybe Bool

Whether TimeZone is automatically derived from the geolocation attribute.

message :: Maybe Text

If an error occurred, an informational message about the error.

geolocationFormat :: Maybe Text

The format of the geolocation attribute. Valid Values:"LAT_LONG" and "CC_POSTALCODE".

timeZone :: Maybe Text

The single time zone applied to every item in the dataset

lastModificationTime :: Maybe POSIX

The last time the resource was modified. The timestamp depends on the status of the job:

CREATE_PENDING - The CreationTime.

CREATE_IN_PROGRESS - The current timestamp.

CREATE_STOPPING - The current timestamp.

CREATE_STOPPED - When the job stopped.

ACTIVE or CREATE_FAILED - When the job finished or failed.

httpStatus :: Int

The response's http status code.

newDescribeDatasetImportJobResponse

Arguments

:: Int
$sel:httpStatus:DescribeDatasetImportJobResponse'

-> DescribeDatasetImportJobResponse

Create a value of DescribeDatasetImportJobResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:creationTime:DescribeDatasetImportJobResponse', describeDatasetImportJobResponse_creationTime - When the dataset import job was created.

$sel:status:DescribeDatasetImportJobResponse', describeDatasetImportJobResponse_status - The status of the dataset import job. States include:

ACTIVE

CREATE_PENDING, CREATE_IN_PROGRESS, CREATE_FAILED

DELETE_PENDING, DELETE_IN_PROGRESS, DELETE_FAILED

CREATE_STOPPING, CREATE_STOPPED

$sel:datasetImportJobName:DescribeDatasetImportJobResponse', describeDatasetImportJobResponse_datasetImportJobName - The name of the dataset import job.

$sel:datasetArn:DescribeDatasetImportJobResponse', describeDatasetImportJobResponse_datasetArn - The Amazon Resource Name (ARN) of the dataset that the training data was imported to.

$sel:timestampFormat:DescribeDatasetImportJobResponse', describeDatasetImportJobResponse_timestampFormat - The format of timestamps in the dataset. The format that you specify depends on the DataFrequency specified when the dataset was created. The following formats are supported

"yyyy-MM-dd"

For the following data frequencies: Y, M, W, and D

"yyyy-MM-dd HH:mm:ss"

For the following data frequencies: H, 30min, 15min, and 1min; and optionally, for: Y, M, W, and D

$sel:dataSize:DescribeDatasetImportJobResponse', describeDatasetImportJobResponse_dataSize - The size of the dataset in gigabytes (GB) after the import job has finished.

$sel:estimatedTimeRemainingInMinutes:DescribeDatasetImportJobResponse', describeDatasetImportJobResponse_estimatedTimeRemainingInMinutes - The estimated time remaining in minutes for the dataset import job to complete.

$sel:fieldStatistics:DescribeDatasetImportJobResponse', describeDatasetImportJobResponse_fieldStatistics - Statistical information about each field in the input data.

$sel:dataSource:DescribeDatasetImportJobResponse', describeDatasetImportJobResponse_dataSource - The location of the training data to import and an AWS Identity and Access Management (IAM) role that Amazon Forecast can assume to access the data.

If encryption is used, DataSource includes an AWS Key Management Service (KMS) key.

$sel:datasetImportJobArn:DescribeDatasetImportJob', describeDatasetImportJobResponse_datasetImportJobArn - The ARN of the dataset import job.

$sel:useGeolocationForTimeZone:DescribeDatasetImportJobResponse', describeDatasetImportJobResponse_useGeolocationForTimeZone - Whether TimeZone is automatically derived from the geolocation attribute.

$sel:message:DescribeDatasetImportJobResponse', describeDatasetImportJobResponse_message - If an error occurred, an informational message about the error.

$sel:geolocationFormat:DescribeDatasetImportJobResponse', describeDatasetImportJobResponse_geolocationFormat - The format of the geolocation attribute. Valid Values:"LAT_LONG" and "CC_POSTALCODE".

$sel:timeZone:DescribeDatasetImportJobResponse', describeDatasetImportJobResponse_timeZone - The single time zone applied to every item in the dataset

$sel:lastModificationTime:DescribeDatasetImportJobResponse', describeDatasetImportJobResponse_lastModificationTime - The last time the resource was modified. The timestamp depends on the status of the job:

CREATE_PENDING - The CreationTime.

CREATE_IN_PROGRESS - The current timestamp.

CREATE_STOPPING - The current timestamp.

CREATE_STOPPED - When the job stopped.

ACTIVE or CREATE_FAILED - When the job finished or failed.

$sel:httpStatus:DescribeDatasetImportJobResponse', describeDatasetImportJobResponse_httpStatus - The response's http status code.

Response Lenses

describeDatasetImportJobResponse_creationTime :: Lens' DescribeDatasetImportJobResponse (Maybe UTCTime)

When the dataset import job was created.

describeDatasetImportJobResponse_status :: Lens' DescribeDatasetImportJobResponse (Maybe Text)

The status of the dataset import job. States include:

ACTIVE

CREATE_PENDING, CREATE_IN_PROGRESS, CREATE_FAILED

DELETE_PENDING, DELETE_IN_PROGRESS, DELETE_FAILED

CREATE_STOPPING, CREATE_STOPPED

describeDatasetImportJobResponse_datasetImportJobName :: Lens' DescribeDatasetImportJobResponse (Maybe Text)

The name of the dataset import job.

describeDatasetImportJobResponse_datasetArn :: Lens' DescribeDatasetImportJobResponse (Maybe Text)

The Amazon Resource Name (ARN) of the dataset that the training data was imported to.

describeDatasetImportJobResponse_timestampFormat :: Lens' DescribeDatasetImportJobResponse (Maybe Text)

The format of timestamps in the dataset. The format that you specify depends on the DataFrequency specified when the dataset was created. The following formats are supported

"yyyy-MM-dd"

For the following data frequencies: Y, M, W, and D

"yyyy-MM-dd HH:mm:ss"

For the following data frequencies: H, 30min, 15min, and 1min; and optionally, for: Y, M, W, and D

describeDatasetImportJobResponse_dataSize :: Lens' DescribeDatasetImportJobResponse (Maybe Double)

The size of the dataset in gigabytes (GB) after the import job has finished.

describeDatasetImportJobResponse_estimatedTimeRemainingInMinutes :: Lens' DescribeDatasetImportJobResponse (Maybe Integer)

The estimated time remaining in minutes for the dataset import job to complete.

describeDatasetImportJobResponse_fieldStatistics :: Lens' DescribeDatasetImportJobResponse (Maybe (HashMap Text Statistics))

Statistical information about each field in the input data.

describeDatasetImportJobResponse_dataSource :: Lens' DescribeDatasetImportJobResponse (Maybe DataSource)

The location of the training data to import and an AWS Identity and Access Management (IAM) role that Amazon Forecast can assume to access the data.

If encryption is used, DataSource includes an AWS Key Management Service (KMS) key.

describeDatasetImportJobResponse_datasetImportJobArn :: Lens' DescribeDatasetImportJobResponse (Maybe Text)

The ARN of the dataset import job.

describeDatasetImportJobResponse_useGeolocationForTimeZone :: Lens' DescribeDatasetImportJobResponse (Maybe Bool)

Whether TimeZone is automatically derived from the geolocation attribute.

describeDatasetImportJobResponse_message :: Lens' DescribeDatasetImportJobResponse (Maybe Text)

If an error occurred, an informational message about the error.

describeDatasetImportJobResponse_geolocationFormat :: Lens' DescribeDatasetImportJobResponse (Maybe Text)

The format of the geolocation attribute. Valid Values:"LAT_LONG" and "CC_POSTALCODE".

describeDatasetImportJobResponse_timeZone :: Lens' DescribeDatasetImportJobResponse (Maybe Text)

The single time zone applied to every item in the dataset

describeDatasetImportJobResponse_lastModificationTime :: Lens' DescribeDatasetImportJobResponse (Maybe UTCTime)

The last time the resource was modified. The timestamp depends on the status of the job:

CREATE_PENDING - The CreationTime.

CREATE_IN_PROGRESS - The current timestamp.

CREATE_STOPPING - The current timestamp.

CREATE_STOPPED - When the job stopped.

ACTIVE or CREATE_FAILED - When the job finished or failed.

describeDatasetImportJobResponse_httpStatus :: Lens' DescribeDatasetImportJobResponse Int

The response's http status code.