Amazonka.Forecast.DescribePredictor

Description

Describes a predictor created using the CreatePredictor operation.

In addition to listing the properties provided in the CreatePredictor request, this operation lists the following properties:

DatasetImportJobArns - The dataset import jobs used to import training data.

AutoMLAlgorithmArns - If AutoML is performed, the algorithms that were evaluated.

CreationTime

LastModificationTime

Status

Message - If an error occurred, information about the error.

Creating a Request

data DescribePredictor

See: newDescribePredictor smart constructor.

Constructors

DescribePredictor'

Fields

predictorArn :: Text

The Amazon Resource Name (ARN) of the predictor that you want information about.

newDescribePredictor

Arguments

:: Text
$sel:predictorArn:DescribePredictor'

-> DescribePredictor

Create a value of DescribePredictor with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:predictorArn:DescribePredictor', describePredictor_predictorArn - The Amazon Resource Name (ARN) of the predictor that you want information about.

Request Lenses

describePredictor_predictorArn :: Lens' DescribePredictor Text

The Amazon Resource Name (ARN) of the predictor that you want information about.

Destructuring the Response

data DescribePredictorResponse

See: newDescribePredictorResponse smart constructor.

Constructors

DescribePredictorResponse'

Fields

creationTime :: Maybe POSIX

When the model training task was created.

forecastHorizon :: Maybe Int

The number of time-steps of the forecast. The forecast horizon is also called the prediction length.

status :: Maybe Text

The status of the predictor. States include:

ACTIVE

CREATE_PENDING, CREATE_IN_PROGRESS, CREATE_FAILED

DELETE_PENDING, DELETE_IN_PROGRESS, DELETE_FAILED

CREATE_STOPPING, CREATE_STOPPED

The Status of the predictor must be ACTIVE before you can use the predictor to create a forecast.

performAutoML :: Maybe Bool

Whether the predictor is set to perform AutoML.

autoMLAlgorithmArns :: Maybe [Text]

When PerformAutoML is specified, the ARN of the chosen algorithm.

trainingParameters :: Maybe (HashMap Text Text)

The default training parameters or overrides selected during model training. When running AutoML or choosing HPO with CNN-QR or DeepAR+, the optimized values for the chosen hyperparameters are returned. For more information, see aws-forecast-choosing-recipes.

algorithmArn :: Maybe Text

The Amazon Resource Name (ARN) of the algorithm used for model training.

hPOConfig :: Maybe HyperParameterTuningJobConfig

The hyperparameter override values for the algorithm.

predictorArn :: Maybe Text

The ARN of the predictor.

optimizationMetric :: Maybe OptimizationMetric

The accuracy metric used to optimize the predictor.

predictorExecutionDetails :: Maybe PredictorExecutionDetails

Details on the the status and results of the backtests performed to evaluate the accuracy of the predictor. You specify the number of backtests to perform when you call the operation.

datasetImportJobArns :: Maybe [Text]

An array of the ARNs of the dataset import jobs used to import training data for the predictor.

estimatedTimeRemainingInMinutes :: Maybe Integer

The estimated time remaining in minutes for the predictor training job to complete.

autoMLOverrideStrategy :: Maybe AutoMLOverrideStrategy

The LatencyOptimized AutoML override strategy is only available in private beta. Contact AWS Support or your account manager to learn more about access privileges.

The AutoML strategy used to train the predictor. Unless LatencyOptimized is specified, the AutoML strategy optimizes predictor accuracy.

This parameter is only valid for predictors trained using AutoML.

evaluationParameters :: Maybe EvaluationParameters

Used to override the default evaluation parameters of the specified algorithm. Amazon Forecast evaluates a predictor by splitting a dataset into training data and testing data. The evaluation parameters define how to perform the split and the number of iterations.

inputDataConfig :: Maybe InputDataConfig

Describes the dataset group that contains the data to use to train the predictor.

predictorName :: Maybe Text

The name of the predictor.

featurizationConfig :: Maybe FeaturizationConfig

The featurization configuration.

encryptionConfig :: Maybe EncryptionConfig

An AWS Key Management Service (KMS) key and the AWS Identity and Access Management (IAM) role that Amazon Forecast can assume to access the key.

forecastTypes :: Maybe (NonEmpty Text)

The forecast types used during predictor training. Default value is ["0.1","0.5","0.9"]

message :: Maybe Text

If an error occurred, an informational message about the error.

performHPO :: Maybe Bool

Whether the predictor is set to perform hyperparameter optimization (HPO).

lastModificationTime :: Maybe POSIX

The last time the resource was modified. The timestamp depends on the status of the job:

CREATE_PENDING - The CreationTime.

CREATE_IN_PROGRESS - The current timestamp.

CREATE_STOPPING - The current timestamp.

CREATE_STOPPED - When the job stopped.

ACTIVE or CREATE_FAILED - When the job finished or failed.

httpStatus :: Int

The response's http status code.

newDescribePredictorResponse

Arguments

:: Int
$sel:httpStatus:DescribePredictorResponse'

-> DescribePredictorResponse

Create a value of DescribePredictorResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:creationTime:DescribePredictorResponse', describePredictorResponse_creationTime - When the model training task was created.

$sel:forecastHorizon:DescribePredictorResponse', describePredictorResponse_forecastHorizon - The number of time-steps of the forecast. The forecast horizon is also called the prediction length.

$sel:status:DescribePredictorResponse', describePredictorResponse_status - The status of the predictor. States include:

ACTIVE

CREATE_PENDING, CREATE_IN_PROGRESS, CREATE_FAILED

DELETE_PENDING, DELETE_IN_PROGRESS, DELETE_FAILED

CREATE_STOPPING, CREATE_STOPPED

The Status of the predictor must be ACTIVE before you can use the predictor to create a forecast.

$sel:performAutoML:DescribePredictorResponse', describePredictorResponse_performAutoML - Whether the predictor is set to perform AutoML.

$sel:autoMLAlgorithmArns:DescribePredictorResponse', describePredictorResponse_autoMLAlgorithmArns - When PerformAutoML is specified, the ARN of the chosen algorithm.

$sel:trainingParameters:DescribePredictorResponse', describePredictorResponse_trainingParameters - The default training parameters or overrides selected during model training. When running AutoML or choosing HPO with CNN-QR or DeepAR+, the optimized values for the chosen hyperparameters are returned. For more information, see aws-forecast-choosing-recipes.

$sel:algorithmArn:DescribePredictorResponse', describePredictorResponse_algorithmArn - The Amazon Resource Name (ARN) of the algorithm used for model training.

$sel:hPOConfig:DescribePredictorResponse', describePredictorResponse_hPOConfig - The hyperparameter override values for the algorithm.

$sel:predictorArn:DescribePredictor', describePredictorResponse_predictorArn - The ARN of the predictor.

$sel:optimizationMetric:DescribePredictorResponse', describePredictorResponse_optimizationMetric - The accuracy metric used to optimize the predictor.

$sel:predictorExecutionDetails:DescribePredictorResponse', describePredictorResponse_predictorExecutionDetails - Details on the the status and results of the backtests performed to evaluate the accuracy of the predictor. You specify the number of backtests to perform when you call the operation.

$sel:datasetImportJobArns:DescribePredictorResponse', describePredictorResponse_datasetImportJobArns - An array of the ARNs of the dataset import jobs used to import training data for the predictor.

$sel:estimatedTimeRemainingInMinutes:DescribePredictorResponse', describePredictorResponse_estimatedTimeRemainingInMinutes - The estimated time remaining in minutes for the predictor training job to complete.

$sel:autoMLOverrideStrategy:DescribePredictorResponse', describePredictorResponse_autoMLOverrideStrategy - The LatencyOptimized AutoML override strategy is only available in private beta. Contact AWS Support or your account manager to learn more about access privileges.

The AutoML strategy used to train the predictor. Unless LatencyOptimized is specified, the AutoML strategy optimizes predictor accuracy.

This parameter is only valid for predictors trained using AutoML.

$sel:evaluationParameters:DescribePredictorResponse', describePredictorResponse_evaluationParameters - Used to override the default evaluation parameters of the specified algorithm. Amazon Forecast evaluates a predictor by splitting a dataset into training data and testing data. The evaluation parameters define how to perform the split and the number of iterations.

$sel:inputDataConfig:DescribePredictorResponse', describePredictorResponse_inputDataConfig - Describes the dataset group that contains the data to use to train the predictor.

$sel:predictorName:DescribePredictorResponse', describePredictorResponse_predictorName - The name of the predictor.

$sel:featurizationConfig:DescribePredictorResponse', describePredictorResponse_featurizationConfig - The featurization configuration.

$sel:encryptionConfig:DescribePredictorResponse', describePredictorResponse_encryptionConfig - An AWS Key Management Service (KMS) key and the AWS Identity and Access Management (IAM) role that Amazon Forecast can assume to access the key.

$sel:forecastTypes:DescribePredictorResponse', describePredictorResponse_forecastTypes - The forecast types used during predictor training. Default value is ["0.1","0.5","0.9"]

$sel:message:DescribePredictorResponse', describePredictorResponse_message - If an error occurred, an informational message about the error.

$sel:performHPO:DescribePredictorResponse', describePredictorResponse_performHPO - Whether the predictor is set to perform hyperparameter optimization (HPO).

$sel:lastModificationTime:DescribePredictorResponse', describePredictorResponse_lastModificationTime - The last time the resource was modified. The timestamp depends on the status of the job:

CREATE_PENDING - The CreationTime.

CREATE_IN_PROGRESS - The current timestamp.

CREATE_STOPPING - The current timestamp.

CREATE_STOPPED - When the job stopped.

ACTIVE or CREATE_FAILED - When the job finished or failed.

$sel:httpStatus:DescribePredictorResponse', describePredictorResponse_httpStatus - The response's http status code.

Response Lenses

describePredictorResponse_creationTime :: Lens' DescribePredictorResponse (Maybe UTCTime)

When the model training task was created.

describePredictorResponse_forecastHorizon :: Lens' DescribePredictorResponse (Maybe Int)

The number of time-steps of the forecast. The forecast horizon is also called the prediction length.

describePredictorResponse_status :: Lens' DescribePredictorResponse (Maybe Text)

The status of the predictor. States include:

ACTIVE

CREATE_PENDING, CREATE_IN_PROGRESS, CREATE_FAILED

DELETE_PENDING, DELETE_IN_PROGRESS, DELETE_FAILED

CREATE_STOPPING, CREATE_STOPPED

The Status of the predictor must be ACTIVE before you can use the predictor to create a forecast.

describePredictorResponse_performAutoML :: Lens' DescribePredictorResponse (Maybe Bool)

Whether the predictor is set to perform AutoML.

describePredictorResponse_autoMLAlgorithmArns :: Lens' DescribePredictorResponse (Maybe [Text])

When PerformAutoML is specified, the ARN of the chosen algorithm.

describePredictorResponse_trainingParameters :: Lens' DescribePredictorResponse (Maybe (HashMap Text Text))

The default training parameters or overrides selected during model training. When running AutoML or choosing HPO with CNN-QR or DeepAR+, the optimized values for the chosen hyperparameters are returned. For more information, see aws-forecast-choosing-recipes.

describePredictorResponse_algorithmArn :: Lens' DescribePredictorResponse (Maybe Text)

The Amazon Resource Name (ARN) of the algorithm used for model training.

describePredictorResponse_hPOConfig :: Lens' DescribePredictorResponse (Maybe HyperParameterTuningJobConfig)

The hyperparameter override values for the algorithm.

describePredictorResponse_predictorArn :: Lens' DescribePredictorResponse (Maybe Text)

The ARN of the predictor.

describePredictorResponse_optimizationMetric :: Lens' DescribePredictorResponse (Maybe OptimizationMetric)

The accuracy metric used to optimize the predictor.

describePredictorResponse_predictorExecutionDetails :: Lens' DescribePredictorResponse (Maybe PredictorExecutionDetails)

Details on the the status and results of the backtests performed to evaluate the accuracy of the predictor. You specify the number of backtests to perform when you call the operation.

describePredictorResponse_datasetImportJobArns :: Lens' DescribePredictorResponse (Maybe [Text])

An array of the ARNs of the dataset import jobs used to import training data for the predictor.

describePredictorResponse_estimatedTimeRemainingInMinutes :: Lens' DescribePredictorResponse (Maybe Integer)

The estimated time remaining in minutes for the predictor training job to complete.

describePredictorResponse_autoMLOverrideStrategy :: Lens' DescribePredictorResponse (Maybe AutoMLOverrideStrategy)

The LatencyOptimized AutoML override strategy is only available in private beta. Contact AWS Support or your account manager to learn more about access privileges.

The AutoML strategy used to train the predictor. Unless LatencyOptimized is specified, the AutoML strategy optimizes predictor accuracy.

This parameter is only valid for predictors trained using AutoML.

describePredictorResponse_evaluationParameters :: Lens' DescribePredictorResponse (Maybe EvaluationParameters)

Used to override the default evaluation parameters of the specified algorithm. Amazon Forecast evaluates a predictor by splitting a dataset into training data and testing data. The evaluation parameters define how to perform the split and the number of iterations.

describePredictorResponse_inputDataConfig :: Lens' DescribePredictorResponse (Maybe InputDataConfig)

Describes the dataset group that contains the data to use to train the predictor.

describePredictorResponse_predictorName :: Lens' DescribePredictorResponse (Maybe Text)

The name of the predictor.

describePredictorResponse_featurizationConfig :: Lens' DescribePredictorResponse (Maybe FeaturizationConfig)

The featurization configuration.

describePredictorResponse_encryptionConfig :: Lens' DescribePredictorResponse (Maybe EncryptionConfig)

An AWS Key Management Service (KMS) key and the AWS Identity and Access Management (IAM) role that Amazon Forecast can assume to access the key.

describePredictorResponse_forecastTypes :: Lens' DescribePredictorResponse (Maybe (NonEmpty Text))

The forecast types used during predictor training. Default value is ["0.1","0.5","0.9"]

describePredictorResponse_message :: Lens' DescribePredictorResponse (Maybe Text)

If an error occurred, an informational message about the error.

describePredictorResponse_performHPO :: Lens' DescribePredictorResponse (Maybe Bool)

Whether the predictor is set to perform hyperparameter optimization (HPO).

describePredictorResponse_lastModificationTime :: Lens' DescribePredictorResponse (Maybe UTCTime)

The last time the resource was modified. The timestamp depends on the status of the job:

CREATE_PENDING - The CreationTime.

CREATE_IN_PROGRESS - The current timestamp.

CREATE_STOPPING - The current timestamp.

CREATE_STOPPED - When the job stopped.

ACTIVE or CREATE_FAILED - When the job finished or failed.

describePredictorResponse_httpStatus :: Lens' DescribePredictorResponse Int

The response's http status code.