Amazonka.Forecast.DescribePredictorBacktestExportJob

Description

Describes a predictor backtest export job created using the CreatePredictorBacktestExportJob operation.

In addition to listing the properties provided by the user in the CreatePredictorBacktestExportJob request, this operation lists the following properties:

CreationTime

LastModificationTime

Status

Message (if an error occurred)

Creating a Request

data DescribePredictorBacktestExportJob

See: newDescribePredictorBacktestExportJob smart constructor.

Constructors

DescribePredictorBacktestExportJob'

Fields

predictorBacktestExportJobArn :: Text

The Amazon Resource Name (ARN) of the predictor backtest export job.

newDescribePredictorBacktestExportJob

Arguments

:: Text
$sel:predictorBacktestExportJobArn:DescribePredictorBacktestExportJob'

-> DescribePredictorBacktestExportJob

Create a value of DescribePredictorBacktestExportJob with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:predictorBacktestExportJobArn:DescribePredictorBacktestExportJob', describePredictorBacktestExportJob_predictorBacktestExportJobArn - The Amazon Resource Name (ARN) of the predictor backtest export job.

Request Lenses

describePredictorBacktestExportJob_predictorBacktestExportJobArn :: Lens' DescribePredictorBacktestExportJob Text

The Amazon Resource Name (ARN) of the predictor backtest export job.

Destructuring the Response

data DescribePredictorBacktestExportJobResponse

See: newDescribePredictorBacktestExportJobResponse smart constructor.

Constructors

DescribePredictorBacktestExportJobResponse'

Fields

creationTime :: Maybe POSIX

When the predictor backtest export job was created.

status :: Maybe Text

The status of the predictor backtest export job. States include:

ACTIVE

CREATE_PENDING, CREATE_IN_PROGRESS, CREATE_FAILED

CREATE_STOPPING, CREATE_STOPPED

DELETE_PENDING, DELETE_IN_PROGRESS, DELETE_FAILED

destination :: Maybe DataDestination

predictorArn :: Maybe Text

The Amazon Resource Name (ARN) of the predictor.

predictorBacktestExportJobArn :: Maybe Text

The Amazon Resource Name (ARN) of the predictor backtest export job.

message :: Maybe Text

Information about any errors that may have occurred during the backtest export.

predictorBacktestExportJobName :: Maybe Text

The name of the predictor backtest export job.

lastModificationTime :: Maybe POSIX

The last time the resource was modified. The timestamp depends on the status of the job:

CREATE_PENDING - The CreationTime.

CREATE_IN_PROGRESS - The current timestamp.

CREATE_STOPPING - The current timestamp.

CREATE_STOPPED - When the job stopped.

ACTIVE or CREATE_FAILED - When the job finished or failed.

httpStatus :: Int

The response's http status code.

newDescribePredictorBacktestExportJobResponse

Arguments

:: Int
$sel:httpStatus:DescribePredictorBacktestExportJobResponse'

-> DescribePredictorBacktestExportJobResponse

Create a value of DescribePredictorBacktestExportJobResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:creationTime:DescribePredictorBacktestExportJobResponse', describePredictorBacktestExportJobResponse_creationTime - When the predictor backtest export job was created.

$sel:status:DescribePredictorBacktestExportJobResponse', describePredictorBacktestExportJobResponse_status - The status of the predictor backtest export job. States include:

ACTIVE

CREATE_PENDING, CREATE_IN_PROGRESS, CREATE_FAILED

CREATE_STOPPING, CREATE_STOPPED

DELETE_PENDING, DELETE_IN_PROGRESS, DELETE_FAILED

$sel:destination:DescribePredictorBacktestExportJobResponse', describePredictorBacktestExportJobResponse_destination - Undocumented member.

$sel:predictorArn:DescribePredictorBacktestExportJobResponse', describePredictorBacktestExportJobResponse_predictorArn - The Amazon Resource Name (ARN) of the predictor.

$sel:predictorBacktestExportJobArn:DescribePredictorBacktestExportJob', describePredictorBacktestExportJobResponse_predictorBacktestExportJobArn - The Amazon Resource Name (ARN) of the predictor backtest export job.

$sel:message:DescribePredictorBacktestExportJobResponse', describePredictorBacktestExportJobResponse_message - Information about any errors that may have occurred during the backtest export.

$sel:predictorBacktestExportJobName:DescribePredictorBacktestExportJobResponse', describePredictorBacktestExportJobResponse_predictorBacktestExportJobName - The name of the predictor backtest export job.

$sel:lastModificationTime:DescribePredictorBacktestExportJobResponse', describePredictorBacktestExportJobResponse_lastModificationTime - The last time the resource was modified. The timestamp depends on the status of the job:

CREATE_PENDING - The CreationTime.

CREATE_IN_PROGRESS - The current timestamp.

CREATE_STOPPING - The current timestamp.

CREATE_STOPPED - When the job stopped.

ACTIVE or CREATE_FAILED - When the job finished or failed.

$sel:httpStatus:DescribePredictorBacktestExportJobResponse', describePredictorBacktestExportJobResponse_httpStatus - The response's http status code.

Response Lenses

describePredictorBacktestExportJobResponse_creationTime :: Lens' DescribePredictorBacktestExportJobResponse (Maybe UTCTime)

When the predictor backtest export job was created.

describePredictorBacktestExportJobResponse_status :: Lens' DescribePredictorBacktestExportJobResponse (Maybe Text)

The status of the predictor backtest export job. States include:

ACTIVE

CREATE_PENDING, CREATE_IN_PROGRESS, CREATE_FAILED

CREATE_STOPPING, CREATE_STOPPED

DELETE_PENDING, DELETE_IN_PROGRESS, DELETE_FAILED

describePredictorBacktestExportJobResponse_destination :: Lens' DescribePredictorBacktestExportJobResponse (Maybe DataDestination)

Undocumented member.

describePredictorBacktestExportJobResponse_predictorArn :: Lens' DescribePredictorBacktestExportJobResponse (Maybe Text)

The Amazon Resource Name (ARN) of the predictor.

describePredictorBacktestExportJobResponse_predictorBacktestExportJobArn :: Lens' DescribePredictorBacktestExportJobResponse (Maybe Text)

The Amazon Resource Name (ARN) of the predictor backtest export job.

describePredictorBacktestExportJobResponse_message :: Lens' DescribePredictorBacktestExportJobResponse (Maybe Text)

Information about any errors that may have occurred during the backtest export.

describePredictorBacktestExportJobResponse_predictorBacktestExportJobName :: Lens' DescribePredictorBacktestExportJobResponse (Maybe Text)

The name of the predictor backtest export job.

describePredictorBacktestExportJobResponse_lastModificationTime :: Lens' DescribePredictorBacktestExportJobResponse (Maybe UTCTime)

The last time the resource was modified. The timestamp depends on the status of the job:

CREATE_PENDING - The CreationTime.

CREATE_IN_PROGRESS - The current timestamp.

CREATE_STOPPING - The current timestamp.

CREATE_STOPPED - When the job stopped.

ACTIVE or CREATE_FAILED - When the job finished or failed.

describePredictorBacktestExportJobResponse_httpStatus :: Lens' DescribePredictorBacktestExportJobResponse Int

The response's http status code.