Amazonka.Forecast.GetAccuracyMetrics

Description

Provides metrics on the accuracy of the models that were trained by the CreatePredictor operation. Use metrics to see how well the model performed and to decide whether to use the predictor to generate a forecast. For more information, see Predictor Metrics.

This operation generates metrics for each backtest window that was evaluated. The number of backtest windows (NumberOfBacktestWindows) is specified using the EvaluationParameters object, which is optionally included in the CreatePredictor request. If NumberOfBacktestWindows isn't specified, the number defaults to one.

The parameters of the filling method determine which items contribute to the metrics. If you want all items to contribute, specify zero. If you want only those items that have complete data in the range being evaluated to contribute, specify nan. For more information, see FeaturizationMethod.

Before you can get accuracy metrics, the Status of the predictor must be ACTIVE, signifying that training has completed. To get the status, use the DescribePredictor operation.

Creating a Request

data GetAccuracyMetrics

See: newGetAccuracyMetrics smart constructor.

Constructors

GetAccuracyMetrics'

Fields

predictorArn :: Text

The Amazon Resource Name (ARN) of the predictor to get metrics for.

newGetAccuracyMetrics

Arguments

:: Text
$sel:predictorArn:GetAccuracyMetrics'

-> GetAccuracyMetrics

Create a value of GetAccuracyMetrics with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:predictorArn:GetAccuracyMetrics', getAccuracyMetrics_predictorArn - The Amazon Resource Name (ARN) of the predictor to get metrics for.

Request Lenses

getAccuracyMetrics_predictorArn :: Lens' GetAccuracyMetrics Text

The Amazon Resource Name (ARN) of the predictor to get metrics for.

Destructuring the Response

data GetAccuracyMetricsResponse

See: newGetAccuracyMetricsResponse smart constructor.

Constructors

GetAccuracyMetricsResponse'

Fields

predictorEvaluationResults :: Maybe [EvaluationResult]

An array of results from evaluating the predictor.

optimizationMetric :: Maybe OptimizationMetric

The accuracy metric used to optimize the predictor.

autoMLOverrideStrategy :: Maybe AutoMLOverrideStrategy

The LatencyOptimized AutoML override strategy is only available in private beta. Contact AWS Support or your account manager to learn more about access privileges.

The AutoML strategy used to train the predictor. Unless LatencyOptimized is specified, the AutoML strategy optimizes predictor accuracy.

This parameter is only valid for predictors trained using AutoML.

httpStatus :: Int

The response's http status code.

newGetAccuracyMetricsResponse

Arguments

:: Int
$sel:httpStatus:GetAccuracyMetricsResponse'

-> GetAccuracyMetricsResponse

Create a value of GetAccuracyMetricsResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:predictorEvaluationResults:GetAccuracyMetricsResponse', getAccuracyMetricsResponse_predictorEvaluationResults - An array of results from evaluating the predictor.

$sel:optimizationMetric:GetAccuracyMetricsResponse', getAccuracyMetricsResponse_optimizationMetric - The accuracy metric used to optimize the predictor.

$sel:autoMLOverrideStrategy:GetAccuracyMetricsResponse', getAccuracyMetricsResponse_autoMLOverrideStrategy - The LatencyOptimized AutoML override strategy is only available in private beta. Contact AWS Support or your account manager to learn more about access privileges.

The AutoML strategy used to train the predictor. Unless LatencyOptimized is specified, the AutoML strategy optimizes predictor accuracy.

This parameter is only valid for predictors trained using AutoML.

$sel:httpStatus:GetAccuracyMetricsResponse', getAccuracyMetricsResponse_httpStatus - The response's http status code.

Response Lenses

getAccuracyMetricsResponse_predictorEvaluationResults :: Lens' GetAccuracyMetricsResponse (Maybe [EvaluationResult])

An array of results from evaluating the predictor.

getAccuracyMetricsResponse_optimizationMetric :: Lens' GetAccuracyMetricsResponse (Maybe OptimizationMetric)

The accuracy metric used to optimize the predictor.

getAccuracyMetricsResponse_autoMLOverrideStrategy :: Lens' GetAccuracyMetricsResponse (Maybe AutoMLOverrideStrategy)

The LatencyOptimized AutoML override strategy is only available in private beta. Contact AWS Support or your account manager to learn more about access privileges.

The AutoML strategy used to train the predictor. Unless LatencyOptimized is specified, the AutoML strategy optimizes predictor accuracy.

This parameter is only valid for predictors trained using AutoML.

getAccuracyMetricsResponse_httpStatus :: Lens' GetAccuracyMetricsResponse Int

The response's http status code.