Amazonka.Forecast.ListDatasetGroups

Description

Returns a list of dataset groups created using the CreateDatasetGroup operation. For each dataset group, this operation returns a summary of its properties, including its Amazon Resource Name (ARN). You can retrieve the complete set of properties by using the dataset group ARN with the DescribeDatasetGroup operation.

This operation returns paginated results.

Creating a Request

data ListDatasetGroups

See: newListDatasetGroups smart constructor.

Constructors

ListDatasetGroups'

Fields

nextToken :: Maybe Text

If the result of the previous request was truncated, the response includes a NextToken. To retrieve the next set of results, use the token in the next request. Tokens expire after 24 hours.

maxResults :: Maybe Natural

The number of items to return in the response.

newListDatasetGroups :: ListDatasetGroups

Create a value of ListDatasetGroups with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:nextToken:ListDatasetGroups', listDatasetGroups_nextToken - If the result of the previous request was truncated, the response includes a NextToken. To retrieve the next set of results, use the token in the next request. Tokens expire after 24 hours.

$sel:maxResults:ListDatasetGroups', listDatasetGroups_maxResults - The number of items to return in the response.

Request Lenses

listDatasetGroups_nextToken :: Lens' ListDatasetGroups (Maybe Text)

If the result of the previous request was truncated, the response includes a NextToken. To retrieve the next set of results, use the token in the next request. Tokens expire after 24 hours.

listDatasetGroups_maxResults :: Lens' ListDatasetGroups (Maybe Natural)

The number of items to return in the response.

Destructuring the Response

data ListDatasetGroupsResponse

See: newListDatasetGroupsResponse smart constructor.

Constructors

ListDatasetGroupsResponse'

Fields

nextToken :: Maybe Text

If the response is truncated, Amazon Forecast returns this token. To retrieve the next set of results, use the token in the next request.

datasetGroups :: Maybe [DatasetGroupSummary]

An array of objects that summarize each dataset group's properties.

httpStatus :: Int

The response's http status code.

newListDatasetGroupsResponse

Arguments

:: Int
$sel:httpStatus:ListDatasetGroupsResponse'

-> ListDatasetGroupsResponse

Create a value of ListDatasetGroupsResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:nextToken:ListDatasetGroups', listDatasetGroupsResponse_nextToken - If the response is truncated, Amazon Forecast returns this token. To retrieve the next set of results, use the token in the next request.

$sel:datasetGroups:ListDatasetGroupsResponse', listDatasetGroupsResponse_datasetGroups - An array of objects that summarize each dataset group's properties.

$sel:httpStatus:ListDatasetGroupsResponse', listDatasetGroupsResponse_httpStatus - The response's http status code.

Response Lenses

listDatasetGroupsResponse_nextToken :: Lens' ListDatasetGroupsResponse (Maybe Text)

If the response is truncated, Amazon Forecast returns this token. To retrieve the next set of results, use the token in the next request.

listDatasetGroupsResponse_datasetGroups :: Lens' ListDatasetGroupsResponse (Maybe [DatasetGroupSummary])

An array of objects that summarize each dataset group's properties.

listDatasetGroupsResponse_httpStatus :: Lens' ListDatasetGroupsResponse Int

The response's http status code.