Amazonka.Forecast.Types.DatasetSummary

Description

Documentation

data DatasetSummary

Provides a summary of the dataset properties used in the ListDatasets operation. To get the complete set of properties, call the DescribeDataset operation, and provide the DatasetArn.

See: newDatasetSummary smart constructor.

Constructors

DatasetSummary'

Fields

creationTime :: Maybe POSIX

When the dataset was created.

datasetArn :: Maybe Text

The Amazon Resource Name (ARN) of the dataset.

domain :: Maybe Domain

The domain associated with the dataset.

datasetType :: Maybe DatasetType

The dataset type.

datasetName :: Maybe Text

The name of the dataset.

lastModificationTime :: Maybe POSIX

When you create a dataset, LastModificationTime is the same as CreationTime. While data is being imported to the dataset, LastModificationTime is the current time of the ListDatasets call. After a CreateDatasetImportJob operation has finished, LastModificationTime is when the import job completed or failed.

newDatasetSummary :: DatasetSummary

Create a value of DatasetSummary with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:creationTime:DatasetSummary', datasetSummary_creationTime - When the dataset was created.

$sel:datasetArn:DatasetSummary', datasetSummary_datasetArn - The Amazon Resource Name (ARN) of the dataset.

$sel:domain:DatasetSummary', datasetSummary_domain - The domain associated with the dataset.

$sel:datasetType:DatasetSummary', datasetSummary_datasetType - The dataset type.

$sel:datasetName:DatasetSummary', datasetSummary_datasetName - The name of the dataset.

$sel:lastModificationTime:DatasetSummary', datasetSummary_lastModificationTime - When you create a dataset, LastModificationTime is the same as CreationTime. While data is being imported to the dataset, LastModificationTime is the current time of the ListDatasets call. After a CreateDatasetImportJob operation has finished, LastModificationTime is when the import job completed or failed.

datasetSummary_creationTime :: Lens' DatasetSummary (Maybe UTCTime)

When the dataset was created.

datasetSummary_datasetArn :: Lens' DatasetSummary (Maybe Text)

The Amazon Resource Name (ARN) of the dataset.

datasetSummary_domain :: Lens' DatasetSummary (Maybe Domain)

The domain associated with the dataset.

datasetSummary_datasetType :: Lens' DatasetSummary (Maybe DatasetType)

The dataset type.

datasetSummary_datasetName :: Lens' DatasetSummary (Maybe Text)

The name of the dataset.

datasetSummary_lastModificationTime :: Lens' DatasetSummary (Maybe UTCTime)

When you create a dataset, LastModificationTime is the same as CreationTime. While data is being imported to the dataset, LastModificationTime is the current time of the ListDatasets call. After a CreateDatasetImportJob operation has finished, LastModificationTime is when the import job completed or failed.