Amazonka.Forecast.Types.Featurization

Description

Documentation

data Featurization

Provides featurization (transformation) information for a dataset field. This object is part of the FeaturizationConfig object.

For example:

{

"AttributeName": "demand",

FeaturizationPipeline [ {

"FeaturizationMethodName": "filling",

"FeaturizationMethodParameters": {"aggregation": "avg", "backfill": "nan"}

} ]

}

See: newFeaturization smart constructor.

Constructors

Featurization'

Fields

featurizationPipeline :: Maybe (NonEmpty FeaturizationMethod)

An array of one FeaturizationMethod object that specifies the feature transformation method.

attributeName :: Text

The name of the schema attribute that specifies the data field to be featurized. Amazon Forecast supports the target field of the TARGET_TIME_SERIES and the RELATED_TIME_SERIES datasets. For example, for the RETAIL domain, the target is demand, and for the CUSTOM domain, the target is target_value. For more information, see howitworks-missing-values.

newFeaturization

Arguments

:: Text
$sel:attributeName:Featurization'

-> Featurization

Create a value of Featurization with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:featurizationPipeline:Featurization', featurization_featurizationPipeline - An array of one FeaturizationMethod object that specifies the feature transformation method.

$sel:attributeName:Featurization', featurization_attributeName - The name of the schema attribute that specifies the data field to be featurized. Amazon Forecast supports the target field of the TARGET_TIME_SERIES and the RELATED_TIME_SERIES datasets. For example, for the RETAIL domain, the target is demand, and for the CUSTOM domain, the target is target_value. For more information, see howitworks-missing-values.

featurization_featurizationPipeline :: Lens' Featurization (Maybe (NonEmpty FeaturizationMethod))

An array of one FeaturizationMethod object that specifies the feature transformation method.

featurization_attributeName :: Lens' Featurization Text

The name of the schema attribute that specifies the data field to be featurized. Amazon Forecast supports the target field of the TARGET_TIME_SERIES and the RELATED_TIME_SERIES datasets. For example, for the RETAIL domain, the target is demand, and for the CUSTOM domain, the target is target_value. For more information, see howitworks-missing-values.