Amazonka.Forecast.Types.ForecastSummary

Description

Documentation

data ForecastSummary

Provides a summary of the forecast properties used in the ListForecasts operation. To get the complete set of properties, call the DescribeForecast operation, and provide the ForecastArn that is listed in the summary.

See: newForecastSummary smart constructor.

Constructors

ForecastSummary'

Fields

creationTime :: Maybe POSIX

When the forecast creation task was created.

status :: Maybe Text

The status of the forecast. States include:

ACTIVE

CREATE_PENDING, CREATE_IN_PROGRESS, CREATE_FAILED

CREATE_STOPPING, CREATE_STOPPED

DELETE_PENDING, DELETE_IN_PROGRESS, DELETE_FAILED

The Status of the forecast must be ACTIVE before you can query or export the forecast.

predictorArn :: Maybe Text

The ARN of the predictor used to generate the forecast.

forecastArn :: Maybe Text

The ARN of the forecast.

forecastName :: Maybe Text

The name of the forecast.

datasetGroupArn :: Maybe Text

The Amazon Resource Name (ARN) of the dataset group that provided the data used to train the predictor.

message :: Maybe Text

If an error occurred, an informational message about the error.

lastModificationTime :: Maybe POSIX

The last time the resource was modified. The timestamp depends on the status of the job:

CREATE_PENDING - The CreationTime.

CREATE_IN_PROGRESS - The current timestamp.

CREATE_STOPPING - The current timestamp.

CREATE_STOPPED - When the job stopped.

ACTIVE or CREATE_FAILED - When the job finished or failed.

newForecastSummary :: ForecastSummary

Create a value of ForecastSummary with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:creationTime:ForecastSummary', forecastSummary_creationTime - When the forecast creation task was created.

$sel:status:ForecastSummary', forecastSummary_status - The status of the forecast. States include:

ACTIVE

CREATE_PENDING, CREATE_IN_PROGRESS, CREATE_FAILED

CREATE_STOPPING, CREATE_STOPPED

DELETE_PENDING, DELETE_IN_PROGRESS, DELETE_FAILED

The Status of the forecast must be ACTIVE before you can query or export the forecast.

$sel:predictorArn:ForecastSummary', forecastSummary_predictorArn - The ARN of the predictor used to generate the forecast.

$sel:forecastArn:ForecastSummary', forecastSummary_forecastArn - The ARN of the forecast.

$sel:forecastName:ForecastSummary', forecastSummary_forecastName - The name of the forecast.

$sel:datasetGroupArn:ForecastSummary', forecastSummary_datasetGroupArn - The Amazon Resource Name (ARN) of the dataset group that provided the data used to train the predictor.

$sel:message:ForecastSummary', forecastSummary_message - If an error occurred, an informational message about the error.

$sel:lastModificationTime:ForecastSummary', forecastSummary_lastModificationTime - The last time the resource was modified. The timestamp depends on the status of the job:

CREATE_PENDING - The CreationTime.

CREATE_IN_PROGRESS - The current timestamp.

CREATE_STOPPING - The current timestamp.

CREATE_STOPPED - When the job stopped.

ACTIVE or CREATE_FAILED - When the job finished or failed.

forecastSummary_creationTime :: Lens' ForecastSummary (Maybe UTCTime)

When the forecast creation task was created.

forecastSummary_status :: Lens' ForecastSummary (Maybe Text)

The status of the forecast. States include:

ACTIVE

CREATE_PENDING, CREATE_IN_PROGRESS, CREATE_FAILED

CREATE_STOPPING, CREATE_STOPPED

DELETE_PENDING, DELETE_IN_PROGRESS, DELETE_FAILED

The Status of the forecast must be ACTIVE before you can query or export the forecast.

forecastSummary_predictorArn :: Lens' ForecastSummary (Maybe Text)

The ARN of the predictor used to generate the forecast.

forecastSummary_forecastArn :: Lens' ForecastSummary (Maybe Text)

The ARN of the forecast.

forecastSummary_forecastName :: Lens' ForecastSummary (Maybe Text)

The name of the forecast.

forecastSummary_datasetGroupArn :: Lens' ForecastSummary (Maybe Text)

The Amazon Resource Name (ARN) of the dataset group that provided the data used to train the predictor.

forecastSummary_message :: Lens' ForecastSummary (Maybe Text)

If an error occurred, an informational message about the error.

forecastSummary_lastModificationTime :: Lens' ForecastSummary (Maybe UTCTime)

The last time the resource was modified. The timestamp depends on the status of the job:

CREATE_PENDING - The CreationTime.

CREATE_IN_PROGRESS - The current timestamp.

CREATE_STOPPING - The current timestamp.

CREATE_STOPPED - When the job stopped.

ACTIVE or CREATE_FAILED - When the job finished or failed.