Amazonka.Forecast.Types.PredictorSummary

Description

Documentation

data PredictorSummary

Provides a summary of the predictor properties that are used in the ListPredictors operation. To get the complete set of properties, call the DescribePredictor operation, and provide the listed PredictorArn.

See: newPredictorSummary smart constructor.

Constructors

PredictorSummary'

Fields

creationTime :: Maybe POSIX

When the model training task was created.

status :: Maybe Text

The status of the predictor. States include:

ACTIVE

CREATE_PENDING, CREATE_IN_PROGRESS, CREATE_FAILED

DELETE_PENDING, DELETE_IN_PROGRESS, DELETE_FAILED

CREATE_STOPPING, CREATE_STOPPED

The Status of the predictor must be ACTIVE before you can use the predictor to create a forecast.

predictorArn :: Maybe Text

The ARN of the predictor.

predictorName :: Maybe Text

The name of the predictor.

datasetGroupArn :: Maybe Text

The Amazon Resource Name (ARN) of the dataset group that contains the data used to train the predictor.

message :: Maybe Text

If an error occurred, an informational message about the error.

lastModificationTime :: Maybe POSIX

The last time the resource was modified. The timestamp depends on the status of the job:

CREATE_PENDING - The CreationTime.

CREATE_IN_PROGRESS - The current timestamp.

CREATE_STOPPING - The current timestamp.

CREATE_STOPPED - When the job stopped.

ACTIVE or CREATE_FAILED - When the job finished or failed.

newPredictorSummary :: PredictorSummary

Create a value of PredictorSummary with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:creationTime:PredictorSummary', predictorSummary_creationTime - When the model training task was created.

$sel:status:PredictorSummary', predictorSummary_status - The status of the predictor. States include:

ACTIVE

CREATE_PENDING, CREATE_IN_PROGRESS, CREATE_FAILED

DELETE_PENDING, DELETE_IN_PROGRESS, DELETE_FAILED

CREATE_STOPPING, CREATE_STOPPED

The Status of the predictor must be ACTIVE before you can use the predictor to create a forecast.

$sel:predictorArn:PredictorSummary', predictorSummary_predictorArn - The ARN of the predictor.

$sel:predictorName:PredictorSummary', predictorSummary_predictorName - The name of the predictor.

$sel:datasetGroupArn:PredictorSummary', predictorSummary_datasetGroupArn - The Amazon Resource Name (ARN) of the dataset group that contains the data used to train the predictor.

$sel:message:PredictorSummary', predictorSummary_message - If an error occurred, an informational message about the error.

$sel:lastModificationTime:PredictorSummary', predictorSummary_lastModificationTime - The last time the resource was modified. The timestamp depends on the status of the job:

CREATE_PENDING - The CreationTime.

CREATE_IN_PROGRESS - The current timestamp.

CREATE_STOPPING - The current timestamp.

CREATE_STOPPED - When the job stopped.

ACTIVE or CREATE_FAILED - When the job finished or failed.

predictorSummary_creationTime :: Lens' PredictorSummary (Maybe UTCTime)

When the model training task was created.

predictorSummary_status :: Lens' PredictorSummary (Maybe Text)

The status of the predictor. States include:

ACTIVE

CREATE_PENDING, CREATE_IN_PROGRESS, CREATE_FAILED

DELETE_PENDING, DELETE_IN_PROGRESS, DELETE_FAILED

CREATE_STOPPING, CREATE_STOPPED

The Status of the predictor must be ACTIVE before you can use the predictor to create a forecast.

predictorSummary_predictorArn :: Lens' PredictorSummary (Maybe Text)

The ARN of the predictor.

predictorSummary_predictorName :: Lens' PredictorSummary (Maybe Text)

The name of the predictor.

predictorSummary_datasetGroupArn :: Lens' PredictorSummary (Maybe Text)

The Amazon Resource Name (ARN) of the dataset group that contains the data used to train the predictor.

predictorSummary_message :: Lens' PredictorSummary (Maybe Text)

If an error occurred, an informational message about the error.

predictorSummary_lastModificationTime :: Lens' PredictorSummary (Maybe UTCTime)

The last time the resource was modified. The timestamp depends on the status of the job:

CREATE_PENDING - The CreationTime.

CREATE_IN_PROGRESS - The current timestamp.

CREATE_STOPPING - The current timestamp.

CREATE_STOPPED - When the job stopped.

ACTIVE or CREATE_FAILED - When the job finished or failed.