Amazonka.Forecast.UpdateDatasetGroup

Description

Replaces the datasets in a dataset group with the specified datasets.

The Status of the dataset group must be ACTIVE before you can use the dataset group to create a predictor. Use the DescribeDatasetGroup operation to get the status.

Creating a Request

data UpdateDatasetGroup

See: newUpdateDatasetGroup smart constructor.

Constructors

UpdateDatasetGroup'

Fields

datasetGroupArn :: Text

The ARN of the dataset group.

datasetArns :: [Text]

An array of the Amazon Resource Names (ARNs) of the datasets to add to the dataset group.

newUpdateDatasetGroup

Arguments

:: Text
$sel:datasetGroupArn:UpdateDatasetGroup'

-> UpdateDatasetGroup

Create a value of UpdateDatasetGroup with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:datasetGroupArn:UpdateDatasetGroup', updateDatasetGroup_datasetGroupArn - The ARN of the dataset group.

$sel:datasetArns:UpdateDatasetGroup', updateDatasetGroup_datasetArns - An array of the Amazon Resource Names (ARNs) of the datasets to add to the dataset group.

Request Lenses

updateDatasetGroup_datasetGroupArn :: Lens' UpdateDatasetGroup Text

The ARN of the dataset group.

updateDatasetGroup_datasetArns :: Lens' UpdateDatasetGroup [Text]

An array of the Amazon Resource Names (ARNs) of the datasets to add to the dataset group.

Destructuring the Response

data UpdateDatasetGroupResponse

See: newUpdateDatasetGroupResponse smart constructor.

Constructors

UpdateDatasetGroupResponse'

Fields

httpStatus :: Int

The response's http status code.

newUpdateDatasetGroupResponse

Arguments

:: Int
$sel:httpStatus:UpdateDatasetGroupResponse'

-> UpdateDatasetGroupResponse

Create a value of UpdateDatasetGroupResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:httpStatus:UpdateDatasetGroupResponse', updateDatasetGroupResponse_httpStatus - The response's http status code.

Response Lenses

updateDatasetGroupResponse_httpStatus :: Lens' UpdateDatasetGroupResponse Int

The response's http status code.