Amazonka.FraudDetector.CreateRule

Description

Creates a rule for use with the specified detector.

Creating a Request

data CreateRule

See: newCreateRule smart constructor.

Constructors

CreateRule'

Fields

description :: Maybe Text

The rule description.

tags :: Maybe [Tag]

A collection of key and value pairs.

ruleId :: Text

The rule ID.

detectorId :: Text

The detector ID for the rule's parent detector.

expression :: Sensitive Text

The rule expression.

language :: Language

The language of the rule.

outcomes :: NonEmpty Text

The outcome or outcomes returned when the rule expression matches.

newCreateRule

Arguments

:: Text
$sel:ruleId:CreateRule'

-> Text
$sel:detectorId:CreateRule'

-> Text
$sel:expression:CreateRule'

-> Language
$sel:language:CreateRule'

-> NonEmpty Text
$sel:outcomes:CreateRule'

-> CreateRule

Create a value of CreateRule with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:description:CreateRule', createRule_description - The rule description.

$sel:tags:CreateRule', createRule_tags - A collection of key and value pairs.

$sel:ruleId:CreateRule', createRule_ruleId - The rule ID.

$sel:detectorId:CreateRule', createRule_detectorId - The detector ID for the rule's parent detector.

$sel:expression:CreateRule', createRule_expression - The rule expression.

$sel:language:CreateRule', createRule_language - The language of the rule.

$sel:outcomes:CreateRule', createRule_outcomes - The outcome or outcomes returned when the rule expression matches.

Request Lenses

createRule_description :: Lens' CreateRule (Maybe Text)

The rule description.

createRule_tags :: Lens' CreateRule (Maybe [Tag])

A collection of key and value pairs.

createRule_ruleId :: Lens' CreateRule Text

The rule ID.

createRule_detectorId :: Lens' CreateRule Text

The detector ID for the rule's parent detector.

createRule_expression :: Lens' CreateRule Text

The rule expression.

createRule_language :: Lens' CreateRule Language

The language of the rule.

createRule_outcomes :: Lens' CreateRule (NonEmpty Text)

The outcome or outcomes returned when the rule expression matches.

Destructuring the Response

data CreateRuleResponse

See: newCreateRuleResponse smart constructor.

Constructors

CreateRuleResponse'

Fields

rule :: Maybe Rule

The created rule.

httpStatus :: Int

The response's http status code.

newCreateRuleResponse

Arguments

:: Int
$sel:httpStatus:CreateRuleResponse'

-> CreateRuleResponse

Create a value of CreateRuleResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:rule:CreateRuleResponse', createRuleResponse_rule - The created rule.

$sel:httpStatus:CreateRuleResponse', createRuleResponse_httpStatus - The response's http status code.

Response Lenses

createRuleResponse_rule :: Lens' CreateRuleResponse (Maybe Rule)

The created rule.

createRuleResponse_httpStatus :: Lens' CreateRuleResponse Int

The response's http status code.