Amazonka.FraudDetector.CreateVariable

Description

Creates a variable.

Creating a Request

data CreateVariable

See: newCreateVariable smart constructor.

Constructors

CreateVariable'

Fields

variableType :: Maybe Text

The variable type. For more information see Variable types.

Valid Values: AUTH_CODE | AVS | BILLING_ADDRESS_L1 | BILLING_ADDRESS_L2 | BILLING_CITY | BILLING_COUNTRY | BILLING_NAME | BILLING_PHONE | BILLING_STATE | BILLING_ZIP | CARD_BIN | CATEGORICAL | CURRENCY_CODE | EMAIL_ADDRESS | FINGERPRINT | FRAUD_LABEL | FREE_FORM_TEXT | IP_ADDRESS | NUMERIC | ORDER_ID | PAYMENT_TYPE | PHONE_NUMBER | PRICE | PRODUCT_CATEGORY | SHIPPING_ADDRESS_L1 | SHIPPING_ADDRESS_L2 | SHIPPING_CITY | SHIPPING_COUNTRY | SHIPPING_NAME | SHIPPING_PHONE | SHIPPING_STATE | SHIPPING_ZIP | USERAGENT

description :: Maybe Text

The description.

tags :: Maybe [Tag]

A collection of key and value pairs.

name :: Text

The name of the variable.

dataType :: DataType

The data type.

dataSource :: DataSource

The source of the data.

defaultValue :: Text

The default value for the variable when no value is received.

newCreateVariable

Arguments

:: Text
$sel:name:CreateVariable'

-> DataType
$sel:dataType:CreateVariable'

-> DataSource
$sel:dataSource:CreateVariable'

-> Text
$sel:defaultValue:CreateVariable'

-> CreateVariable

Create a value of CreateVariable with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:variableType:CreateVariable', createVariable_variableType - The variable type. For more information see Variable types.

Valid Values: AUTH_CODE | AVS | BILLING_ADDRESS_L1 | BILLING_ADDRESS_L2 | BILLING_CITY | BILLING_COUNTRY | BILLING_NAME | BILLING_PHONE | BILLING_STATE | BILLING_ZIP | CARD_BIN | CATEGORICAL | CURRENCY_CODE | EMAIL_ADDRESS | FINGERPRINT | FRAUD_LABEL | FREE_FORM_TEXT | IP_ADDRESS | NUMERIC | ORDER_ID | PAYMENT_TYPE | PHONE_NUMBER | PRICE | PRODUCT_CATEGORY | SHIPPING_ADDRESS_L1 | SHIPPING_ADDRESS_L2 | SHIPPING_CITY | SHIPPING_COUNTRY | SHIPPING_NAME | SHIPPING_PHONE | SHIPPING_STATE | SHIPPING_ZIP | USERAGENT

$sel:description:CreateVariable', createVariable_description - The description.

$sel:tags:CreateVariable', createVariable_tags - A collection of key and value pairs.

$sel:name:CreateVariable', createVariable_name - The name of the variable.

$sel:dataType:CreateVariable', createVariable_dataType - The data type.

$sel:dataSource:CreateVariable', createVariable_dataSource - The source of the data.

$sel:defaultValue:CreateVariable', createVariable_defaultValue - The default value for the variable when no value is received.

Request Lenses

createVariable_variableType :: Lens' CreateVariable (Maybe Text)

The variable type. For more information see Variable types.

Valid Values: AUTH_CODE | AVS | BILLING_ADDRESS_L1 | BILLING_ADDRESS_L2 | BILLING_CITY | BILLING_COUNTRY | BILLING_NAME | BILLING_PHONE | BILLING_STATE | BILLING_ZIP | CARD_BIN | CATEGORICAL | CURRENCY_CODE | EMAIL_ADDRESS | FINGERPRINT | FRAUD_LABEL | FREE_FORM_TEXT | IP_ADDRESS | NUMERIC | ORDER_ID | PAYMENT_TYPE | PHONE_NUMBER | PRICE | PRODUCT_CATEGORY | SHIPPING_ADDRESS_L1 | SHIPPING_ADDRESS_L2 | SHIPPING_CITY | SHIPPING_COUNTRY | SHIPPING_NAME | SHIPPING_PHONE | SHIPPING_STATE | SHIPPING_ZIP | USERAGENT

createVariable_description :: Lens' CreateVariable (Maybe Text)

The description.

createVariable_tags :: Lens' CreateVariable (Maybe [Tag])

A collection of key and value pairs.

createVariable_name :: Lens' CreateVariable Text

The name of the variable.

createVariable_dataType :: Lens' CreateVariable DataType

The data type.

createVariable_dataSource :: Lens' CreateVariable DataSource

The source of the data.

createVariable_defaultValue :: Lens' CreateVariable Text

The default value for the variable when no value is received.

Destructuring the Response

data CreateVariableResponse

See: newCreateVariableResponse smart constructor.

Constructors

CreateVariableResponse'

Fields

httpStatus :: Int

The response's http status code.

newCreateVariableResponse

Arguments

:: Int
$sel:httpStatus:CreateVariableResponse'

-> CreateVariableResponse

Create a value of CreateVariableResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:httpStatus:CreateVariableResponse', createVariableResponse_httpStatus - The response's http status code.

Response Lenses

createVariableResponse_httpStatus :: Lens' CreateVariableResponse Int

The response's http status code.