Amazonka.FraudDetector.DeleteEvent

Description

Deletes the specified event.

When you delete an event, Amazon Fraud Detector permanently deletes that event and the event data is no longer stored in Amazon Fraud Detector.

Creating a Request

data DeleteEvent

See: newDeleteEvent smart constructor.

Constructors

DeleteEvent'

Fields

deleteAuditHistory :: Maybe Bool

Specifies whether or not to delete any predictions associated with the event.

eventId :: Text

The ID of the event to delete.

eventTypeName :: Text

The name of the event type.

newDeleteEvent

Arguments

:: Text
$sel:eventId:DeleteEvent'

-> Text
$sel:eventTypeName:DeleteEvent'

-> DeleteEvent

Create a value of DeleteEvent with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:deleteAuditHistory:DeleteEvent', deleteEvent_deleteAuditHistory - Specifies whether or not to delete any predictions associated with the event.

$sel:eventId:DeleteEvent', deleteEvent_eventId - The ID of the event to delete.

$sel:eventTypeName:DeleteEvent', deleteEvent_eventTypeName - The name of the event type.

Request Lenses

deleteEvent_deleteAuditHistory :: Lens' DeleteEvent (Maybe Bool)

Specifies whether or not to delete any predictions associated with the event.

deleteEvent_eventId :: Lens' DeleteEvent Text

The ID of the event to delete.

deleteEvent_eventTypeName :: Lens' DeleteEvent Text

The name of the event type.

Destructuring the Response

data DeleteEventResponse

See: newDeleteEventResponse smart constructor.

Constructors

DeleteEventResponse'

Fields

httpStatus :: Int

The response's http status code.

newDeleteEventResponse

Arguments

:: Int
$sel:httpStatus:DeleteEventResponse'

-> DeleteEventResponse

Create a value of DeleteEventResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:httpStatus:DeleteEventResponse', deleteEventResponse_httpStatus - The response's http status code.

Response Lenses

deleteEventResponse_httpStatus :: Lens' DeleteEventResponse Int

The response's http status code.