Amazonka.FraudDetector.GetExternalModels

Description

Gets the details for one or more Amazon SageMaker models that have been imported into the service. This is a paginated API. If you provide a null maxResults, this actions retrieves a maximum of 10 records per page. If you provide a maxResults, the value must be between 5 and 10. To get the next page results, provide the pagination token from the GetExternalModelsResult as part of your request. A null pagination token fetches the records from the beginning.

Creating a Request

data GetExternalModels

See: newGetExternalModels smart constructor.

Constructors

GetExternalModels'

Fields

modelEndpoint :: Maybe Text

The Amazon SageMaker model endpoint.

nextToken :: Maybe Text

The next page token for the request.

maxResults :: Maybe Natural

The maximum number of objects to return for the request.

newGetExternalModels :: GetExternalModels

Create a value of GetExternalModels with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:modelEndpoint:GetExternalModels', getExternalModels_modelEndpoint - The Amazon SageMaker model endpoint.

$sel:nextToken:GetExternalModels', getExternalModels_nextToken - The next page token for the request.

$sel:maxResults:GetExternalModels', getExternalModels_maxResults - The maximum number of objects to return for the request.

Request Lenses

getExternalModels_modelEndpoint :: Lens' GetExternalModels (Maybe Text)

The Amazon SageMaker model endpoint.

getExternalModels_nextToken :: Lens' GetExternalModels (Maybe Text)

The next page token for the request.

getExternalModels_maxResults :: Lens' GetExternalModels (Maybe Natural)

The maximum number of objects to return for the request.

Destructuring the Response

data GetExternalModelsResponse

See: newGetExternalModelsResponse smart constructor.

Constructors

GetExternalModelsResponse'

Fields

nextToken :: Maybe Text

The next page token to be used in subsequent requests.

externalModels :: Maybe [ExternalModel]

Gets the Amazon SageMaker models.

httpStatus :: Int

The response's http status code.

newGetExternalModelsResponse

Arguments

:: Int
$sel:httpStatus:GetExternalModelsResponse'

-> GetExternalModelsResponse

Create a value of GetExternalModelsResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:nextToken:GetExternalModels', getExternalModelsResponse_nextToken - The next page token to be used in subsequent requests.

$sel:externalModels:GetExternalModelsResponse', getExternalModelsResponse_externalModels - Gets the Amazon SageMaker models.

$sel:httpStatus:GetExternalModelsResponse', getExternalModelsResponse_httpStatus - The response's http status code.

Response Lenses

getExternalModelsResponse_nextToken :: Lens' GetExternalModelsResponse (Maybe Text)

The next page token to be used in subsequent requests.

getExternalModelsResponse_externalModels :: Lens' GetExternalModelsResponse (Maybe [ExternalModel])

Gets the Amazon SageMaker models.

getExternalModelsResponse_httpStatus :: Lens' GetExternalModelsResponse Int

The response's http status code.