Amazonka.FraudDetector.GetOutcomes

Description

Gets one or more outcomes. This is a paginated API. If you provide a null maxResults, this actions retrieves a maximum of 100 records per page. If you provide a maxResults, the value must be between 50 and 100. To get the next page results, provide the pagination token from the GetOutcomesResult as part of your request. A null pagination token fetches the records from the beginning.

Creating a Request

data GetOutcomes

See: newGetOutcomes smart constructor.

Constructors

GetOutcomes'

Fields

nextToken :: Maybe Text

The next page token for the request.

name :: Maybe Text

The name of the outcome or outcomes to get.

maxResults :: Maybe Natural

The maximum number of objects to return for the request.

newGetOutcomes :: GetOutcomes

Create a value of GetOutcomes with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:nextToken:GetOutcomes', getOutcomes_nextToken - The next page token for the request.

$sel:name:GetOutcomes', getOutcomes_name - The name of the outcome or outcomes to get.

$sel:maxResults:GetOutcomes', getOutcomes_maxResults - The maximum number of objects to return for the request.

Request Lenses

getOutcomes_nextToken :: Lens' GetOutcomes (Maybe Text)

The next page token for the request.

getOutcomes_name :: Lens' GetOutcomes (Maybe Text)

The name of the outcome or outcomes to get.

getOutcomes_maxResults :: Lens' GetOutcomes (Maybe Natural)

The maximum number of objects to return for the request.

Destructuring the Response

data GetOutcomesResponse

See: newGetOutcomesResponse smart constructor.

Constructors

GetOutcomesResponse'

Fields

outcomes :: Maybe [Outcome]

The outcomes.

nextToken :: Maybe Text

The next page token for subsequent requests.

httpStatus :: Int

The response's http status code.

newGetOutcomesResponse

Arguments

:: Int
$sel:httpStatus:GetOutcomesResponse'

-> GetOutcomesResponse

Create a value of GetOutcomesResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:outcomes:GetOutcomesResponse', getOutcomesResponse_outcomes - The outcomes.

$sel:nextToken:GetOutcomes', getOutcomesResponse_nextToken - The next page token for subsequent requests.

$sel:httpStatus:GetOutcomesResponse', getOutcomesResponse_httpStatus - The response's http status code.

Response Lenses

getOutcomesResponse_outcomes :: Lens' GetOutcomesResponse (Maybe [Outcome])

The outcomes.

getOutcomesResponse_nextToken :: Lens' GetOutcomesResponse (Maybe Text)

The next page token for subsequent requests.

getOutcomesResponse_httpStatus :: Lens' GetOutcomesResponse Int

The response's http status code.