Amazonka.FraudDetector.SendEvent

Description

Stores events in Amazon Fraud Detector without generating fraud predictions for those events. For example, you can use SendEvent to upload a historical dataset, which you can then later use to train a model.

Creating a Request

data SendEvent

See: newSendEvent smart constructor.

Constructors

SendEvent'

Fields

labelTimestamp :: Maybe Text

The timestamp associated with the label. Required if specifying assignedLabel.

assignedLabel :: Maybe Text

The label to associate with the event. Required if specifying labelTimestamp.

eventId :: Text

The event ID to upload.

eventTypeName :: Text

The event type name of the event.

eventTimestamp :: Text

The timestamp that defines when the event under evaluation occurred. The timestamp must be specified using ISO 8601 standard in UTC.

eventVariables :: HashMap Text (Sensitive Text)

Names of the event type's variables you defined in Amazon Fraud Detector to represent data elements and their corresponding values for the event you are sending for evaluation.

entities :: [Sensitive Entity]

An array of entities.

newSendEvent

Arguments

:: Text
$sel:eventId:SendEvent'

-> Text
$sel:eventTypeName:SendEvent'

-> Text
$sel:eventTimestamp:SendEvent'

-> SendEvent

Create a value of SendEvent with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:labelTimestamp:SendEvent', sendEvent_labelTimestamp - The timestamp associated with the label. Required if specifying assignedLabel.

$sel:assignedLabel:SendEvent', sendEvent_assignedLabel - The label to associate with the event. Required if specifying labelTimestamp.

$sel:eventId:SendEvent', sendEvent_eventId - The event ID to upload.

$sel:eventTypeName:SendEvent', sendEvent_eventTypeName - The event type name of the event.

$sel:eventTimestamp:SendEvent', sendEvent_eventTimestamp - The timestamp that defines when the event under evaluation occurred. The timestamp must be specified using ISO 8601 standard in UTC.

$sel:eventVariables:SendEvent', sendEvent_eventVariables - Names of the event type's variables you defined in Amazon Fraud Detector to represent data elements and their corresponding values for the event you are sending for evaluation.

$sel:entities:SendEvent', sendEvent_entities - An array of entities.

Request Lenses

sendEvent_labelTimestamp :: Lens' SendEvent (Maybe Text)

The timestamp associated with the label. Required if specifying assignedLabel.

sendEvent_assignedLabel :: Lens' SendEvent (Maybe Text)

The label to associate with the event. Required if specifying labelTimestamp.

sendEvent_eventId :: Lens' SendEvent Text

The event ID to upload.

sendEvent_eventTypeName :: Lens' SendEvent Text

The event type name of the event.

sendEvent_eventTimestamp :: Lens' SendEvent Text

The timestamp that defines when the event under evaluation occurred. The timestamp must be specified using ISO 8601 standard in UTC.

sendEvent_eventVariables :: Lens' SendEvent (HashMap Text Text)

Names of the event type's variables you defined in Amazon Fraud Detector to represent data elements and their corresponding values for the event you are sending for evaluation.

sendEvent_entities :: Lens' SendEvent [Entity]

An array of entities.

Destructuring the Response

data SendEventResponse

See: newSendEventResponse smart constructor.

Constructors

SendEventResponse'

Fields

httpStatus :: Int

The response's http status code.

newSendEventResponse

Arguments

:: Int
$sel:httpStatus:SendEventResponse'

-> SendEventResponse

Create a value of SendEventResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:httpStatus:SendEventResponse', sendEventResponse_httpStatus - The response's http status code.

Response Lenses

sendEventResponse_httpStatus :: Lens' SendEventResponse Int

The response's http status code.