Amazonka.FraudDetector.Types.BatchPrediction

Description

Documentation

data BatchPrediction

The batch prediction details.

See: newBatchPrediction smart constructor.

Constructors

BatchPrediction'

Fields

failureReason :: Maybe Text

The reason a batch prediction job failed.

iamRoleArn :: Maybe Text

The ARN of the IAM role to use for this job request.

status :: Maybe AsyncJobStatus

The batch prediction status.

processedRecordsCount :: Maybe Int

The number of records processed by the batch prediction job.

totalRecordsCount :: Maybe Int

The total number of records in the batch prediction job.

lastHeartbeatTime :: Maybe Text

Timestamp of most recent heartbeat indicating the batch prediction job was making progress.

jobId :: Maybe Text

The job ID for the batch prediction.

arn :: Maybe Text

The ARN of batch prediction job.

startTime :: Maybe Text

Timestamp of when the batch prediction job started.

eventTypeName :: Maybe Text

The name of the event type.

completionTime :: Maybe Text

Timestamp of when the batch prediction job completed.

outputPath :: Maybe Text

The Amazon S3 location of your output file.

detectorName :: Maybe Text

The name of the detector.

detectorVersion :: Maybe Text

The detector version.

inputPath :: Maybe Text

The Amazon S3 location of your training file.

newBatchPrediction :: BatchPrediction

Create a value of BatchPrediction with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:failureReason:BatchPrediction', batchPrediction_failureReason - The reason a batch prediction job failed.

$sel:iamRoleArn:BatchPrediction', batchPrediction_iamRoleArn - The ARN of the IAM role to use for this job request.

$sel:status:BatchPrediction', batchPrediction_status - The batch prediction status.

$sel:processedRecordsCount:BatchPrediction', batchPrediction_processedRecordsCount - The number of records processed by the batch prediction job.

$sel:totalRecordsCount:BatchPrediction', batchPrediction_totalRecordsCount - The total number of records in the batch prediction job.

$sel:lastHeartbeatTime:BatchPrediction', batchPrediction_lastHeartbeatTime - Timestamp of most recent heartbeat indicating the batch prediction job was making progress.

$sel:jobId:BatchPrediction', batchPrediction_jobId - The job ID for the batch prediction.

$sel:arn:BatchPrediction', batchPrediction_arn - The ARN of batch prediction job.

$sel:startTime:BatchPrediction', batchPrediction_startTime - Timestamp of when the batch prediction job started.

$sel:eventTypeName:BatchPrediction', batchPrediction_eventTypeName - The name of the event type.

$sel:completionTime:BatchPrediction', batchPrediction_completionTime - Timestamp of when the batch prediction job completed.

$sel:outputPath:BatchPrediction', batchPrediction_outputPath - The Amazon S3 location of your output file.

$sel:detectorName:BatchPrediction', batchPrediction_detectorName - The name of the detector.

$sel:detectorVersion:BatchPrediction', batchPrediction_detectorVersion - The detector version.

$sel:inputPath:BatchPrediction', batchPrediction_inputPath - The Amazon S3 location of your training file.

batchPrediction_failureReason :: Lens' BatchPrediction (Maybe Text)

The reason a batch prediction job failed.

batchPrediction_iamRoleArn :: Lens' BatchPrediction (Maybe Text)

The ARN of the IAM role to use for this job request.

batchPrediction_status :: Lens' BatchPrediction (Maybe AsyncJobStatus)

The batch prediction status.

batchPrediction_processedRecordsCount :: Lens' BatchPrediction (Maybe Int)

The number of records processed by the batch prediction job.

batchPrediction_totalRecordsCount :: Lens' BatchPrediction (Maybe Int)

The total number of records in the batch prediction job.

batchPrediction_lastHeartbeatTime :: Lens' BatchPrediction (Maybe Text)

Timestamp of most recent heartbeat indicating the batch prediction job was making progress.

batchPrediction_jobId :: Lens' BatchPrediction (Maybe Text)

The job ID for the batch prediction.

batchPrediction_arn :: Lens' BatchPrediction (Maybe Text)

The ARN of batch prediction job.

batchPrediction_startTime :: Lens' BatchPrediction (Maybe Text)

Timestamp of when the batch prediction job started.

batchPrediction_eventTypeName :: Lens' BatchPrediction (Maybe Text)

The name of the event type.

batchPrediction_completionTime :: Lens' BatchPrediction (Maybe Text)

Timestamp of when the batch prediction job completed.

batchPrediction_outputPath :: Lens' BatchPrediction (Maybe Text)

The Amazon S3 location of your output file.

batchPrediction_detectorName :: Lens' BatchPrediction (Maybe Text)

The name of the detector.

batchPrediction_detectorVersion :: Lens' BatchPrediction (Maybe Text)

The detector version.

batchPrediction_inputPath :: Lens' BatchPrediction (Maybe Text)

The Amazon S3 location of your training file.