Amazonka.FraudDetector.Types.Entity

Description

Documentation

data Entity

The entity details.

See: newEntity smart constructor.

Constructors

Entity'

Fields

entityType :: Text

The entity type.

entityId :: Text

The entity ID. If you do not know the entityId, you can pass unknown, which is areserved string literal.

newEntity

Arguments

:: Text
$sel:entityType:Entity'

-> Text
$sel:entityId:Entity'

-> Entity

Create a value of Entity with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:entityType:Entity', entity_entityType - The entity type.

$sel:entityId:Entity', entity_entityId - The entity ID. If you do not know the entityId, you can pass unknown, which is areserved string literal.

entity_entityType :: Lens' Entity Text

The entity type.

entity_entityId :: Lens' Entity Text

The entity ID. If you do not know the entityId, you can pass unknown, which is areserved string literal.