Amazonka.FraudDetector.Types.Event

Description

Documentation

data Event

The event details.

See: newEvent smart constructor.

Constructors

Event'

Fields

eventTimestamp :: Maybe Text

The timestamp that defines when the event under evaluation occurred. The timestamp must be specified using ISO 8601 standard in UTC.

entities :: Maybe [Sensitive Entity]

The event entities.

labelTimestamp :: Maybe Text

The timestamp associated with the label to update. The timestamp must be specified using ISO 8601 standard in UTC.

eventTypeName :: Maybe Text

The event type.

eventVariables :: Maybe (HashMap Text (Sensitive Text))

Names of the event type's variables you defined in Amazon Fraud Detector to represent data elements and their corresponding values for the event you are sending for evaluation.

currentLabel :: Maybe Text

The label associated with the event.

eventId :: Maybe Text

The event ID.

newEvent :: Event

Create a value of Event with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:eventTimestamp:Event', event_eventTimestamp - The timestamp that defines when the event under evaluation occurred. The timestamp must be specified using ISO 8601 standard in UTC.

$sel:entities:Event', event_entities - The event entities.

$sel:labelTimestamp:Event', event_labelTimestamp - The timestamp associated with the label to update. The timestamp must be specified using ISO 8601 standard in UTC.

$sel:eventTypeName:Event', event_eventTypeName - The event type.

$sel:eventVariables:Event', event_eventVariables - Names of the event type's variables you defined in Amazon Fraud Detector to represent data elements and their corresponding values for the event you are sending for evaluation.

$sel:currentLabel:Event', event_currentLabel - The label associated with the event.

$sel:eventId:Event', event_eventId - The event ID.

event_eventTimestamp :: Lens' Event (Maybe Text)

The timestamp that defines when the event under evaluation occurred. The timestamp must be specified using ISO 8601 standard in UTC.

event_entities :: Lens' Event (Maybe [Entity])

The event entities.

event_labelTimestamp :: Lens' Event (Maybe Text)

The timestamp associated with the label to update. The timestamp must be specified using ISO 8601 standard in UTC.

event_eventTypeName :: Lens' Event (Maybe Text)

The event type.

event_eventVariables :: Lens' Event (Maybe (HashMap Text Text))

Names of the event type's variables you defined in Amazon Fraud Detector to represent data elements and their corresponding values for the event you are sending for evaluation.

event_currentLabel :: Lens' Event (Maybe Text)

The label associated with the event.

event_eventId :: Lens' Event (Maybe Text)

The event ID.