Amazonka.FraudDetector.Types.EventType

Description

Documentation

data EventType

The event type details.

See: newEventType smart constructor.

Constructors

EventType'

Fields

lastUpdatedTime :: Maybe Text

Timestamp of when the event type was last updated.

arn :: Maybe Text

The entity type ARN.

createdTime :: Maybe Text

Timestamp of when the event type was created.

entityTypes :: Maybe (NonEmpty Text)

The event type entity types.

eventVariables :: Maybe [Text]

The event type event variables.

name :: Maybe Text

The event type name.

ingestedEventStatistics :: Maybe IngestedEventStatistics

Data about the stored events.

labels :: Maybe [Text]

The event type labels.

eventIngestion :: Maybe EventIngestion

If Enabled, Amazon Fraud Detector stores event data when you generate a prediction and uses that data to update calculated variables in near real-time. Amazon Fraud Detector uses this data, known as INGESTED_EVENTS, to train your model and improve fraud predictions.

description :: Maybe Text

The event type description.

newEventType :: EventType

Create a value of EventType with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:lastUpdatedTime:EventType', eventType_lastUpdatedTime - Timestamp of when the event type was last updated.

$sel:arn:EventType', eventType_arn - The entity type ARN.

$sel:createdTime:EventType', eventType_createdTime - Timestamp of when the event type was created.

$sel:entityTypes:EventType', eventType_entityTypes - The event type entity types.

$sel:eventVariables:EventType', eventType_eventVariables - The event type event variables.

$sel:name:EventType', eventType_name - The event type name.

$sel:ingestedEventStatistics:EventType', eventType_ingestedEventStatistics - Data about the stored events.

$sel:labels:EventType', eventType_labels - The event type labels.

$sel:eventIngestion:EventType', eventType_eventIngestion - If Enabled, Amazon Fraud Detector stores event data when you generate a prediction and uses that data to update calculated variables in near real-time. Amazon Fraud Detector uses this data, known as INGESTED_EVENTS, to train your model and improve fraud predictions.

$sel:description:EventType', eventType_description - The event type description.

eventType_lastUpdatedTime :: Lens' EventType (Maybe Text)

Timestamp of when the event type was last updated.

eventType_arn :: Lens' EventType (Maybe Text)

The entity type ARN.

eventType_createdTime :: Lens' EventType (Maybe Text)

Timestamp of when the event type was created.

eventType_entityTypes :: Lens' EventType (Maybe (NonEmpty Text))

The event type entity types.

eventType_eventVariables :: Lens' EventType (Maybe [Text])

The event type event variables.

eventType_name :: Lens' EventType (Maybe Text)

The event type name.

eventType_ingestedEventStatistics :: Lens' EventType (Maybe IngestedEventStatistics)

Data about the stored events.

eventType_labels :: Lens' EventType (Maybe [Text])

The event type labels.

eventType_eventIngestion :: Lens' EventType (Maybe EventIngestion)

If Enabled, Amazon Fraud Detector stores event data when you generate a prediction and uses that data to update calculated variables in near real-time. Amazon Fraud Detector uses this data, known as INGESTED_EVENTS, to train your model and improve fraud predictions.

eventType_description :: Lens' EventType (Maybe Text)

The event type description.