Amazonka.FraudDetector.Types.IngestedEventsDetail

Description

Documentation

data IngestedEventsDetail

The details of the ingested event.

See: newIngestedEventsDetail smart constructor.

Constructors

IngestedEventsDetail'

Fields

ingestedEventsTimeWindow :: IngestedEventsTimeWindow

The start and stop time of the ingested events.

newIngestedEventsDetail

Arguments

:: IngestedEventsTimeWindow
$sel:ingestedEventsTimeWindow:IngestedEventsDetail'

-> IngestedEventsDetail

Create a value of IngestedEventsDetail with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:ingestedEventsTimeWindow:IngestedEventsDetail', ingestedEventsDetail_ingestedEventsTimeWindow - The start and stop time of the ingested events.

ingestedEventsDetail_ingestedEventsTimeWindow :: Lens' IngestedEventsDetail IngestedEventsTimeWindow

The start and stop time of the ingested events.