Amazonka.FraudDetector.Types.MetricDataPoint

Description

Documentation

data MetricDataPoint

Model performance metrics data points.

See: newMetricDataPoint smart constructor.

Constructors

MetricDataPoint'

Fields

precision :: Maybe Double

The percentage of fraud events correctly predicted as fraudulent as compared to all events predicted as fraudulent.

fpr :: Maybe Double

The false positive rate. This is the percentage of total legitimate events that are incorrectly predicted as fraud.

threshold :: Maybe Double

The model threshold that specifies an acceptable fraud capture rate. For example, a threshold of 500 means any model score 500 or above is labeled as fraud.

tpr :: Maybe Double

The true positive rate. This is the percentage of total fraud the model detects. Also known as capture rate.

newMetricDataPoint :: MetricDataPoint

Create a value of MetricDataPoint with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:precision:MetricDataPoint', metricDataPoint_precision - The percentage of fraud events correctly predicted as fraudulent as compared to all events predicted as fraudulent.

$sel:fpr:MetricDataPoint', metricDataPoint_fpr - The false positive rate. This is the percentage of total legitimate events that are incorrectly predicted as fraud.

$sel:threshold:MetricDataPoint', metricDataPoint_threshold - The model threshold that specifies an acceptable fraud capture rate. For example, a threshold of 500 means any model score 500 or above is labeled as fraud.

$sel:tpr:MetricDataPoint', metricDataPoint_tpr - The true positive rate. This is the percentage of total fraud the model detects. Also known as capture rate.

metricDataPoint_precision :: Lens' MetricDataPoint (Maybe Double)

The percentage of fraud events correctly predicted as fraudulent as compared to all events predicted as fraudulent.

metricDataPoint_fpr :: Lens' MetricDataPoint (Maybe Double)

The false positive rate. This is the percentage of total legitimate events that are incorrectly predicted as fraud.

metricDataPoint_threshold :: Lens' MetricDataPoint (Maybe Double)

The model threshold that specifies an acceptable fraud capture rate. For example, a threshold of 500 means any model score 500 or above is labeled as fraud.

metricDataPoint_tpr :: Lens' MetricDataPoint (Maybe Double)

The true positive rate. This is the percentage of total fraud the model detects. Also known as capture rate.