Amazonka.FraudDetector.Types.ModelVersionDetail

Description

Documentation

data ModelVersionDetail

The details of the model version.

See: newModelVersionDetail smart constructor.

Constructors

ModelVersionDetail'

Fields

status :: Maybe Text

The status of the model version.

modelType :: Maybe ModelTypeEnum

The model type.

lastUpdatedTime :: Maybe Text

The timestamp when the model was last updated.

modelId :: Maybe Text

The model ID.

arn :: Maybe Text

The model version ARN.

trainingDataSource :: Maybe TrainingDataSourceEnum

The model version training data source.

createdTime :: Maybe Text

The timestamp when the model was created.

externalEventsDetail :: Maybe ExternalEventsDetail

The external events data details. This will be populated if the trainingDataSource for the model version is specified as EXTERNAL_EVENTS.

ingestedEventsDetail :: Maybe IngestedEventsDetail

The ingested events data details. This will be populated if the trainingDataSource for the model version is specified as INGESTED_EVENTS.

modelVersionNumber :: Maybe Text

The model version number.

trainingResult :: Maybe TrainingResult

The training results.

trainingDataSchema :: Maybe TrainingDataSchema

The training data schema.

newModelVersionDetail :: ModelVersionDetail

Create a value of ModelVersionDetail with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:status:ModelVersionDetail', modelVersionDetail_status - The status of the model version.

$sel:modelType:ModelVersionDetail', modelVersionDetail_modelType - The model type.

$sel:lastUpdatedTime:ModelVersionDetail', modelVersionDetail_lastUpdatedTime - The timestamp when the model was last updated.

$sel:modelId:ModelVersionDetail', modelVersionDetail_modelId - The model ID.

$sel:arn:ModelVersionDetail', modelVersionDetail_arn - The model version ARN.

$sel:trainingDataSource:ModelVersionDetail', modelVersionDetail_trainingDataSource - The model version training data source.

$sel:createdTime:ModelVersionDetail', modelVersionDetail_createdTime - The timestamp when the model was created.

$sel:externalEventsDetail:ModelVersionDetail', modelVersionDetail_externalEventsDetail - The external events data details. This will be populated if the trainingDataSource for the model version is specified as EXTERNAL_EVENTS.

$sel:ingestedEventsDetail:ModelVersionDetail', modelVersionDetail_ingestedEventsDetail - The ingested events data details. This will be populated if the trainingDataSource for the model version is specified as INGESTED_EVENTS.

$sel:modelVersionNumber:ModelVersionDetail', modelVersionDetail_modelVersionNumber - The model version number.

$sel:trainingResult:ModelVersionDetail', modelVersionDetail_trainingResult - The training results.

$sel:trainingDataSchema:ModelVersionDetail', modelVersionDetail_trainingDataSchema - The training data schema.

modelVersionDetail_status :: Lens' ModelVersionDetail (Maybe Text)

The status of the model version.

modelVersionDetail_modelType :: Lens' ModelVersionDetail (Maybe ModelTypeEnum)

The model type.

modelVersionDetail_lastUpdatedTime :: Lens' ModelVersionDetail (Maybe Text)

The timestamp when the model was last updated.

modelVersionDetail_modelId :: Lens' ModelVersionDetail (Maybe Text)

The model ID.

modelVersionDetail_arn :: Lens' ModelVersionDetail (Maybe Text)

The model version ARN.

modelVersionDetail_trainingDataSource :: Lens' ModelVersionDetail (Maybe TrainingDataSourceEnum)

The model version training data source.

modelVersionDetail_createdTime :: Lens' ModelVersionDetail (Maybe Text)

The timestamp when the model was created.

modelVersionDetail_externalEventsDetail :: Lens' ModelVersionDetail (Maybe ExternalEventsDetail)

The external events data details. This will be populated if the trainingDataSource for the model version is specified as EXTERNAL_EVENTS.

modelVersionDetail_ingestedEventsDetail :: Lens' ModelVersionDetail (Maybe IngestedEventsDetail)

The ingested events data details. This will be populated if the trainingDataSource for the model version is specified as INGESTED_EVENTS.

modelVersionDetail_modelVersionNumber :: Lens' ModelVersionDetail (Maybe Text)

The model version number.

modelVersionDetail_trainingResult :: Lens' ModelVersionDetail (Maybe TrainingResult)

The training results.

modelVersionDetail_trainingDataSchema :: Lens' ModelVersionDetail (Maybe TrainingDataSchema)

The training data schema.