Amazonka.FraudDetector.Types.RuleDetail

Description

Documentation

data RuleDetail

The details of the rule.

See: newRuleDetail smart constructor.

Constructors

RuleDetail'

Fields

ruleVersion :: Maybe Text

The rule version.

lastUpdatedTime :: Maybe Text

Timestamp of the last time the rule was updated.

arn :: Maybe Text

The rule ARN.

createdTime :: Maybe Text

The timestamp of when the rule was created.

ruleId :: Maybe Text

The rule ID.

outcomes :: Maybe (NonEmpty Text)

The rule outcomes.

detectorId :: Maybe Text

The detector for which the rule is associated.

expression :: Maybe (Sensitive Text)

The rule expression.

language :: Maybe Language

The rule language.

description :: Maybe Text

The rule description.

newRuleDetail :: RuleDetail

Create a value of RuleDetail with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:ruleVersion:RuleDetail', ruleDetail_ruleVersion - The rule version.

$sel:lastUpdatedTime:RuleDetail', ruleDetail_lastUpdatedTime - Timestamp of the last time the rule was updated.

$sel:arn:RuleDetail', ruleDetail_arn - The rule ARN.

$sel:createdTime:RuleDetail', ruleDetail_createdTime - The timestamp of when the rule was created.

$sel:ruleId:RuleDetail', ruleDetail_ruleId - The rule ID.

$sel:outcomes:RuleDetail', ruleDetail_outcomes - The rule outcomes.

$sel:detectorId:RuleDetail', ruleDetail_detectorId - The detector for which the rule is associated.

$sel:expression:RuleDetail', ruleDetail_expression - The rule expression.

$sel:language:RuleDetail', ruleDetail_language - The rule language.

$sel:description:RuleDetail', ruleDetail_description - The rule description.

ruleDetail_ruleVersion :: Lens' RuleDetail (Maybe Text)

The rule version.

ruleDetail_lastUpdatedTime :: Lens' RuleDetail (Maybe Text)

Timestamp of the last time the rule was updated.

ruleDetail_arn :: Lens' RuleDetail (Maybe Text)

The rule ARN.

ruleDetail_createdTime :: Lens' RuleDetail (Maybe Text)

The timestamp of when the rule was created.

ruleDetail_ruleId :: Lens' RuleDetail (Maybe Text)

The rule ID.

ruleDetail_outcomes :: Lens' RuleDetail (Maybe (NonEmpty Text))

The rule outcomes.

ruleDetail_detectorId :: Lens' RuleDetail (Maybe Text)

The detector for which the rule is associated.

ruleDetail_expression :: Lens' RuleDetail (Maybe Text)

The rule expression.

ruleDetail_language :: Lens' RuleDetail (Maybe Language)

The rule language.

ruleDetail_description :: Lens' RuleDetail (Maybe Text)

The rule description.