Amazonka.FraudDetector.UpdateModelVersionStatus

Description

Updates the status of a model version.

You can perform the following status updates:

Change the TRAINING_COMPLETE status to ACTIVE.

Change ACTIVE to INACTIVE.

Creating a Request

data UpdateModelVersionStatus

See: newUpdateModelVersionStatus smart constructor.

Constructors

UpdateModelVersionStatus'

Fields

modelId :: Text

The model ID of the model version to update.

modelType :: ModelTypeEnum

The model type.

modelVersionNumber :: Text

The model version number.

status :: ModelVersionStatus

The model version status.

newUpdateModelVersionStatus

Arguments

:: Text
$sel:modelId:UpdateModelVersionStatus'

-> ModelTypeEnum
$sel:modelType:UpdateModelVersionStatus'

-> Text
$sel:modelVersionNumber:UpdateModelVersionStatus'

-> ModelVersionStatus
$sel:status:UpdateModelVersionStatus'

-> UpdateModelVersionStatus

Create a value of UpdateModelVersionStatus with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:modelId:UpdateModelVersionStatus', updateModelVersionStatus_modelId - The model ID of the model version to update.

$sel:modelType:UpdateModelVersionStatus', updateModelVersionStatus_modelType - The model type.

$sel:modelVersionNumber:UpdateModelVersionStatus', updateModelVersionStatus_modelVersionNumber - The model version number.

$sel:status:UpdateModelVersionStatus', updateModelVersionStatus_status - The model version status.

Request Lenses

updateModelVersionStatus_modelId :: Lens' UpdateModelVersionStatus Text

The model ID of the model version to update.

updateModelVersionStatus_modelType :: Lens' UpdateModelVersionStatus ModelTypeEnum

The model type.

updateModelVersionStatus_modelVersionNumber :: Lens' UpdateModelVersionStatus Text

The model version number.

updateModelVersionStatus_status :: Lens' UpdateModelVersionStatus ModelVersionStatus

The model version status.

Destructuring the Response

data UpdateModelVersionStatusResponse

See: newUpdateModelVersionStatusResponse smart constructor.

Constructors

UpdateModelVersionStatusResponse'

Fields

httpStatus :: Int

The response's http status code.

newUpdateModelVersionStatusResponse

Arguments

:: Int
$sel:httpStatus:UpdateModelVersionStatusResponse'

-> UpdateModelVersionStatusResponse

Create a value of UpdateModelVersionStatusResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:httpStatus:UpdateModelVersionStatusResponse', updateModelVersionStatusResponse_httpStatus - The response's http status code.

Response Lenses

updateModelVersionStatusResponse_httpStatus :: Lens' UpdateModelVersionStatusResponse Int

The response's http status code.