Amazonka.GameLift.CreateAlias

Description

Creates an alias for a fleet. In most situations, you can use an alias ID in place of a fleet ID. An alias provides a level of abstraction for a fleet that is useful when redirecting player traffic from one fleet to another, such as when updating your game build.

Amazon GameLift supports two types of routing strategies for aliases: simple and terminal. A simple alias points to an active fleet. A terminal alias is used to display messaging or link to a URL instead of routing players to an active fleet. For example, you might use a terminal alias when a game version is no longer supported and you want to direct players to an upgrade site.

To create a fleet alias, specify an alias name, routing strategy, and optional description. Each simple alias can point to only one fleet, but a fleet can have multiple aliases. If successful, a new alias record is returned, including an alias ID and an ARN. You can reassign an alias to another fleet by calling UpdateAlias.

Related actions

CreateAlias | ListAliases | DescribeAlias | UpdateAlias | DeleteAlias | ResolveAlias | All APIs by task

Creating a Request

data CreateAlias

Represents the input for a request operation.

See: newCreateAlias smart constructor.

Constructors

CreateAlias'

Fields

description :: Maybe Text

A human-readable description of the alias.

tags :: Maybe [Tag]

A list of labels to assign to the new alias resource. Tags are developer-defined key-value pairs. Tagging AWS resources are useful for resource management, access management and cost allocation. For more information, see Tagging AWS Resources in the AWS General Reference. Once the resource is created, you can use TagResource, UntagResource, and ListTagsForResource to add, remove, and view tags. The maximum tag limit may be lower than stated. See the AWS General Reference for actual tagging limits.

name :: Text

A descriptive label that is associated with an alias. Alias names do not need to be unique.

routingStrategy :: RoutingStrategy

The routing configuration, including routing type and fleet target, for the alias.

newCreateAlias

Arguments

:: Text
$sel:name:CreateAlias'

-> RoutingStrategy
$sel:routingStrategy:CreateAlias'

-> CreateAlias

Create a value of CreateAlias with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:description:CreateAlias', createAlias_description - A human-readable description of the alias.

$sel:tags:CreateAlias', createAlias_tags - A list of labels to assign to the new alias resource. Tags are developer-defined key-value pairs. Tagging AWS resources are useful for resource management, access management and cost allocation. For more information, see Tagging AWS Resources in the AWS General Reference. Once the resource is created, you can use TagResource, UntagResource, and ListTagsForResource to add, remove, and view tags. The maximum tag limit may be lower than stated. See the AWS General Reference for actual tagging limits.

$sel:name:CreateAlias', createAlias_name - A descriptive label that is associated with an alias. Alias names do not need to be unique.

$sel:routingStrategy:CreateAlias', createAlias_routingStrategy - The routing configuration, including routing type and fleet target, for the alias.

Request Lenses

createAlias_description :: Lens' CreateAlias (Maybe Text)

A human-readable description of the alias.

createAlias_tags :: Lens' CreateAlias (Maybe [Tag])

A list of labels to assign to the new alias resource. Tags are developer-defined key-value pairs. Tagging AWS resources are useful for resource management, access management and cost allocation. For more information, see Tagging AWS Resources in the AWS General Reference. Once the resource is created, you can use TagResource, UntagResource, and ListTagsForResource to add, remove, and view tags. The maximum tag limit may be lower than stated. See the AWS General Reference for actual tagging limits.

createAlias_name :: Lens' CreateAlias Text

A descriptive label that is associated with an alias. Alias names do not need to be unique.

createAlias_routingStrategy :: Lens' CreateAlias RoutingStrategy

The routing configuration, including routing type and fleet target, for the alias.

Destructuring the Response

data CreateAliasResponse

Represents the returned data in response to a request operation.

See: newCreateAliasResponse smart constructor.

Constructors

CreateAliasResponse'

Fields

alias :: Maybe Alias

The newly created alias resource.

httpStatus :: Int

The response's http status code.

newCreateAliasResponse

Arguments

:: Int
$sel:httpStatus:CreateAliasResponse'

-> CreateAliasResponse

Create a value of CreateAliasResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:alias:CreateAliasResponse', createAliasResponse_alias - The newly created alias resource.

$sel:httpStatus:CreateAliasResponse', createAliasResponse_httpStatus - The response's http status code.

Response Lenses

createAliasResponse_alias :: Lens' CreateAliasResponse (Maybe Alias)

The newly created alias resource.

createAliasResponse_httpStatus :: Lens' CreateAliasResponse Int

The response's http status code.