Amazonka.GameLift.CreateFleet

Description

Creates a fleet of Amazon Elastic Compute Cloud (Amazon EC2) instances to host your custom game server or Realtime Servers. Use this operation to configure the computing resources for your fleet and provide instructions for running game servers on each instance.

Most GameLift fleets can deploy instances to multiple locations, including the home Region (where the fleet is created) and an optional set of remote locations. Fleets that are created in the following AWS Regions support multiple locations: us-east-1 (N. Virginia), us-west-2 (Oregon), eu-central-1 (Frankfurt), eu-west-1 (Ireland), ap-southeast-2 (Sydney), ap-northeast-1 (Tokyo), and ap-northeast-2 (Seoul). Fleets that are created in other GameLift Regions can deploy instances in the fleet's home Region only. All fleet instances use the same configuration regardless of location; however, you can adjust capacity settings and turn auto-scaling on/off for each location.

To create a fleet, choose the hardware for your instances, specify a game server build or Realtime script to deploy, and provide a runtime configuration to direct GameLift how to start and run game servers on each instance in the fleet. Set permissions for inbound traffic to your game servers, and enable optional features as needed. When creating a multi-location fleet, provide a list of additional remote locations.

If successful, this operation creates a new Fleet resource and places it in NEW status, which prompts GameLift to initiate the fleet creation workflow. You can track fleet creation by checking fleet status using DescribeFleetAttributes and DescribeFleetLocationAttributes/, or by monitoring fleet creation events using DescribeFleetEvents. As soon as the fleet status changes to ACTIVE, you can enable automatic scaling for the fleet with PutScalingPolicy and set capacity for the home Region with UpdateFleetCapacity. When the status of each remote location reaches ACTIVE, you can set capacity by location using UpdateFleetCapacity.

Learn more

Setting up fleets

Debug fleet creation issues

Multi-location fleets

Related actions

CreateFleet | UpdateFleetCapacity | PutScalingPolicy | DescribeEC2InstanceLimits | DescribeFleetAttributes | DescribeFleetLocationAttributes | UpdateFleetAttributes | StopFleetActions | DeleteFleet | All APIs by task

Creating a Request

data CreateFleet

Represents the input for a request operation.

See: newCreateFleet smart constructor.

Constructors

CreateFleet'

Fields

serverLaunchParameters :: Maybe Text

This parameter is no longer used. Specify server launch parameters using the RuntimeConfiguration parameter. Requests that use this parameter instead continue to be valid.

logPaths :: Maybe [Text]

This parameter is no longer used. To specify where GameLift should store log files once a server process shuts down, use the GameLift server API ProcessReady() and specify one or more directory paths in logParameters. See more information in the Server API Reference.

peerVpcId :: Maybe Text

A unique identifier for a VPC with resources to be accessed by your GameLift fleet. The VPC must be in the same Region as your fleet. To look up a VPC ID, use the VPC Dashboard in the AWS Management Console. Learn more about VPC peering in VPC Peering with GameLift Fleets.

buildId :: Maybe Text

The unique identifier for a custom game server build to be deployed on fleet instances. You can use either the build ID or ARN. The build must be uploaded to GameLift and in READY status. This fleet property cannot be changed later.

fleetType :: Maybe FleetType

Indicates whether to use On-Demand or Spot instances for this fleet. By default, this property is set to ON_DEMAND. Learn more about when to use On-Demand versus Spot Instances. This property cannot be changed after the fleet is created.

peerVpcAwsAccountId :: Maybe Text

Used when peering your GameLift fleet with a VPC, the unique identifier for the AWS account that owns the VPC. You can find your account ID in the AWS Management Console under account settings.

eC2InboundPermissions :: Maybe [IpPermission]

The allowed IP address ranges and port settings that allow inbound traffic to access game sessions on this fleet. If the fleet is hosting a custom game build, this property must be set before players can connect to game sessions. For Realtime Servers fleets, GameLift automatically sets TCP and UDP ranges.

runtimeConfiguration :: Maybe RuntimeConfiguration

Instructions for how to launch and maintain server processes on instances in the fleet. The runtime configuration defines one or more server process configurations, each identifying a build executable or Realtime script file and the number of processes of that type to run concurrently.

The RuntimeConfiguration parameter is required unless the fleet is being configured using the older parameters ServerLaunchPath and ServerLaunchParameters, which are still supported for backward compatibility.

newGameSessionProtectionPolicy' :: Maybe ProtectionPolicy

The status of termination protection for active game sessions on the fleet. By default, this property is set to NoProtection. You can also set game session protection for an individual game session by calling UpdateGameSession.

NoProtection - Game sessions can be terminated during active gameplay as a result of a scale-down event.

FullProtection - Game sessions in ACTIVE status cannot be terminated during a scale-down event.

scriptId :: Maybe Text

The unique identifier for a Realtime configuration script to be deployed on fleet instances. You can use either the script ID or ARN. Scripts must be uploaded to GameLift prior to creating the fleet. This fleet property cannot be changed later.

certificateConfiguration :: Maybe CertificateConfiguration

Prompts GameLift to generate a TLS/SSL certificate for the fleet. TLS certificates are used for encrypting traffic between game clients and the game servers that are running on GameLift. By default, the CertificateConfiguration is set to DISABLED. Learn more at Securing Client/Server Communication. This property cannot be changed after the fleet is created.

Note: This feature requires the AWS Certificate Manager (ACM) service, which is not available in all AWS regions. When working in a region that does not support this feature, a fleet creation request with certificate generation fails with a 4xx error.

serverLaunchPath :: Maybe Text

This parameter is no longer used. Specify a server launch path using the RuntimeConfiguration parameter. Requests that use this parameter instead continue to be valid.

instanceRoleArn :: Maybe Text

A unique identifier for an AWS IAM role that manages access to your AWS services. With an instance role ARN set, any application that runs on an instance in this fleet can assume the role, including install scripts, server processes, and daemons (background processes). Create a role or look up a role's ARN by using the IAM dashboard in the AWS Management Console. Learn more about using on-box credentials for your game servers at Access external resources from a game server. This property cannot be changed after the fleet is created.

metricGroups :: Maybe [Text]

The name of an AWS CloudWatch metric group to add this fleet to. A metric group is used to aggregate the metrics for multiple fleets. You can specify an existing metric group name or set a new name to create a new metric group. A fleet can be included in only one metric group at a time.

description :: Maybe Text

A human-readable description of the fleet.

resourceCreationLimitPolicy :: Maybe ResourceCreationLimitPolicy

A policy that limits the number of game sessions that an individual player can create on instances in this fleet within a specified span of time.

locations :: Maybe (NonEmpty LocationConfiguration)

A set of remote locations to deploy additional instances to and manage as part of the fleet. This parameter can only be used when creating fleets in AWS Regions that support multiple locations. You can add any GameLift-supported AWS Region as a remote location, in the form of an AWS Region code such as us-west-2. To create a fleet with instances in the home Region only, omit this parameter.

tags :: Maybe [Tag]

A list of labels to assign to the new fleet resource. Tags are developer-defined key-value pairs. Tagging AWS resources are useful for resource management, access management and cost allocation. For more information, see Tagging AWS Resources in the AWS General Reference. Once the fleet is created, you can use TagResource, UntagResource, and ListTagsForResource to add, remove, and view tags. The maximum tag limit may be lower than stated. See the /AWS General Reference/ for actual tagging limits.

name :: Text

A descriptive label that is associated with a fleet. Fleet names do not need to be unique.

eC2InstanceType :: EC2InstanceType

The GameLift-supported EC2 instance type to use for all fleet instances. Instance type determines the computing resources that will be used to host your game servers, including CPU, memory, storage, and networking capacity. See Amazon EC2 Instance Types for detailed descriptions of EC2 instance types.

newCreateFleet

Arguments

:: Text
$sel:name:CreateFleet'

-> EC2InstanceType
$sel:eC2InstanceType:CreateFleet'

-> CreateFleet

Create a value of CreateFleet with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:serverLaunchParameters:CreateFleet', createFleet_serverLaunchParameters - This parameter is no longer used. Specify server launch parameters using the RuntimeConfiguration parameter. Requests that use this parameter instead continue to be valid.

$sel:logPaths:CreateFleet', createFleet_logPaths - This parameter is no longer used. To specify where GameLift should store log files once a server process shuts down, use the GameLift server API ProcessReady() and specify one or more directory paths in logParameters. See more information in the Server API Reference.

$sel:peerVpcId:CreateFleet', createFleet_peerVpcId - A unique identifier for a VPC with resources to be accessed by your GameLift fleet. The VPC must be in the same Region as your fleet. To look up a VPC ID, use the VPC Dashboard in the AWS Management Console. Learn more about VPC peering in VPC Peering with GameLift Fleets.

$sel:buildId:CreateFleet', createFleet_buildId - The unique identifier for a custom game server build to be deployed on fleet instances. You can use either the build ID or ARN. The build must be uploaded to GameLift and in READY status. This fleet property cannot be changed later.

$sel:fleetType:CreateFleet', createFleet_fleetType - Indicates whether to use On-Demand or Spot instances for this fleet. By default, this property is set to ON_DEMAND. Learn more about when to use On-Demand versus Spot Instances. This property cannot be changed after the fleet is created.

$sel:peerVpcAwsAccountId:CreateFleet', createFleet_peerVpcAwsAccountId - Used when peering your GameLift fleet with a VPC, the unique identifier for the AWS account that owns the VPC. You can find your account ID in the AWS Management Console under account settings.

$sel:eC2InboundPermissions:CreateFleet', createFleet_eC2InboundPermissions - The allowed IP address ranges and port settings that allow inbound traffic to access game sessions on this fleet. If the fleet is hosting a custom game build, this property must be set before players can connect to game sessions. For Realtime Servers fleets, GameLift automatically sets TCP and UDP ranges.

$sel:runtimeConfiguration:CreateFleet', createFleet_runtimeConfiguration - Instructions for how to launch and maintain server processes on instances in the fleet. The runtime configuration defines one or more server process configurations, each identifying a build executable or Realtime script file and the number of processes of that type to run concurrently.

The RuntimeConfiguration parameter is required unless the fleet is being configured using the older parameters ServerLaunchPath and ServerLaunchParameters, which are still supported for backward compatibility.

$sel:newGameSessionProtectionPolicy':CreateFleet', createFleet_newGameSessionProtectionPolicy - The status of termination protection for active game sessions on the fleet. By default, this property is set to NoProtection. You can also set game session protection for an individual game session by calling UpdateGameSession.

NoProtection - Game sessions can be terminated during active gameplay as a result of a scale-down event.

FullProtection - Game sessions in ACTIVE status cannot be terminated during a scale-down event.

$sel:scriptId:CreateFleet', createFleet_scriptId - The unique identifier for a Realtime configuration script to be deployed on fleet instances. You can use either the script ID or ARN. Scripts must be uploaded to GameLift prior to creating the fleet. This fleet property cannot be changed later.

$sel:certificateConfiguration:CreateFleet', createFleet_certificateConfiguration - Prompts GameLift to generate a TLS/SSL certificate for the fleet. TLS certificates are used for encrypting traffic between game clients and the game servers that are running on GameLift. By default, the CertificateConfiguration is set to DISABLED. Learn more at Securing Client/Server Communication. This property cannot be changed after the fleet is created.

Note: This feature requires the AWS Certificate Manager (ACM) service, which is not available in all AWS regions. When working in a region that does not support this feature, a fleet creation request with certificate generation fails with a 4xx error.

$sel:serverLaunchPath:CreateFleet', createFleet_serverLaunchPath - This parameter is no longer used. Specify a server launch path using the RuntimeConfiguration parameter. Requests that use this parameter instead continue to be valid.

$sel:instanceRoleArn:CreateFleet', createFleet_instanceRoleArn - A unique identifier for an AWS IAM role that manages access to your AWS services. With an instance role ARN set, any application that runs on an instance in this fleet can assume the role, including install scripts, server processes, and daemons (background processes). Create a role or look up a role's ARN by using the IAM dashboard in the AWS Management Console. Learn more about using on-box credentials for your game servers at Access external resources from a game server. This property cannot be changed after the fleet is created.

$sel:metricGroups:CreateFleet', createFleet_metricGroups - The name of an AWS CloudWatch metric group to add this fleet to. A metric group is used to aggregate the metrics for multiple fleets. You can specify an existing metric group name or set a new name to create a new metric group. A fleet can be included in only one metric group at a time.

$sel:description:CreateFleet', createFleet_description - A human-readable description of the fleet.

$sel:resourceCreationLimitPolicy:CreateFleet', createFleet_resourceCreationLimitPolicy - A policy that limits the number of game sessions that an individual player can create on instances in this fleet within a specified span of time.

$sel:locations:CreateFleet', createFleet_locations - A set of remote locations to deploy additional instances to and manage as part of the fleet. This parameter can only be used when creating fleets in AWS Regions that support multiple locations. You can add any GameLift-supported AWS Region as a remote location, in the form of an AWS Region code such as us-west-2. To create a fleet with instances in the home Region only, omit this parameter.

$sel:tags:CreateFleet', createFleet_tags - A list of labels to assign to the new fleet resource. Tags are developer-defined key-value pairs. Tagging AWS resources are useful for resource management, access management and cost allocation. For more information, see Tagging AWS Resources in the AWS General Reference. Once the fleet is created, you can use TagResource, UntagResource, and ListTagsForResource to add, remove, and view tags. The maximum tag limit may be lower than stated. See the /AWS General Reference/ for actual tagging limits.

$sel:name:CreateFleet', createFleet_name - A descriptive label that is associated with a fleet. Fleet names do not need to be unique.

$sel:eC2InstanceType:CreateFleet', createFleet_eC2InstanceType - The GameLift-supported EC2 instance type to use for all fleet instances. Instance type determines the computing resources that will be used to host your game servers, including CPU, memory, storage, and networking capacity. See Amazon EC2 Instance Types for detailed descriptions of EC2 instance types.

Request Lenses

createFleet_serverLaunchParameters :: Lens' CreateFleet (Maybe Text)

This parameter is no longer used. Specify server launch parameters using the RuntimeConfiguration parameter. Requests that use this parameter instead continue to be valid.

createFleet_logPaths :: Lens' CreateFleet (Maybe [Text])

This parameter is no longer used. To specify where GameLift should store log files once a server process shuts down, use the GameLift server API ProcessReady() and specify one or more directory paths in logParameters. See more information in the Server API Reference.

createFleet_peerVpcId :: Lens' CreateFleet (Maybe Text)

A unique identifier for a VPC with resources to be accessed by your GameLift fleet. The VPC must be in the same Region as your fleet. To look up a VPC ID, use the VPC Dashboard in the AWS Management Console. Learn more about VPC peering in VPC Peering with GameLift Fleets.

createFleet_buildId :: Lens' CreateFleet (Maybe Text)

The unique identifier for a custom game server build to be deployed on fleet instances. You can use either the build ID or ARN. The build must be uploaded to GameLift and in READY status. This fleet property cannot be changed later.

createFleet_fleetType :: Lens' CreateFleet (Maybe FleetType)

Indicates whether to use On-Demand or Spot instances for this fleet. By default, this property is set to ON_DEMAND. Learn more about when to use On-Demand versus Spot Instances. This property cannot be changed after the fleet is created.

createFleet_peerVpcAwsAccountId :: Lens' CreateFleet (Maybe Text)

Used when peering your GameLift fleet with a VPC, the unique identifier for the AWS account that owns the VPC. You can find your account ID in the AWS Management Console under account settings.

createFleet_eC2InboundPermissions :: Lens' CreateFleet (Maybe [IpPermission])

The allowed IP address ranges and port settings that allow inbound traffic to access game sessions on this fleet. If the fleet is hosting a custom game build, this property must be set before players can connect to game sessions. For Realtime Servers fleets, GameLift automatically sets TCP and UDP ranges.

createFleet_runtimeConfiguration :: Lens' CreateFleet (Maybe RuntimeConfiguration)

Instructions for how to launch and maintain server processes on instances in the fleet. The runtime configuration defines one or more server process configurations, each identifying a build executable or Realtime script file and the number of processes of that type to run concurrently.

The RuntimeConfiguration parameter is required unless the fleet is being configured using the older parameters ServerLaunchPath and ServerLaunchParameters, which are still supported for backward compatibility.

createFleet_newGameSessionProtectionPolicy :: Lens' CreateFleet (Maybe ProtectionPolicy)

The status of termination protection for active game sessions on the fleet. By default, this property is set to NoProtection. You can also set game session protection for an individual game session by calling UpdateGameSession.

NoProtection - Game sessions can be terminated during active gameplay as a result of a scale-down event.

FullProtection - Game sessions in ACTIVE status cannot be terminated during a scale-down event.

createFleet_scriptId :: Lens' CreateFleet (Maybe Text)

The unique identifier for a Realtime configuration script to be deployed on fleet instances. You can use either the script ID or ARN. Scripts must be uploaded to GameLift prior to creating the fleet. This fleet property cannot be changed later.

createFleet_certificateConfiguration :: Lens' CreateFleet (Maybe CertificateConfiguration)

Prompts GameLift to generate a TLS/SSL certificate for the fleet. TLS certificates are used for encrypting traffic between game clients and the game servers that are running on GameLift. By default, the CertificateConfiguration is set to DISABLED. Learn more at Securing Client/Server Communication. This property cannot be changed after the fleet is created.

Note: This feature requires the AWS Certificate Manager (ACM) service, which is not available in all AWS regions. When working in a region that does not support this feature, a fleet creation request with certificate generation fails with a 4xx error.

createFleet_serverLaunchPath :: Lens' CreateFleet (Maybe Text)

This parameter is no longer used. Specify a server launch path using the RuntimeConfiguration parameter. Requests that use this parameter instead continue to be valid.

createFleet_instanceRoleArn :: Lens' CreateFleet (Maybe Text)

A unique identifier for an AWS IAM role that manages access to your AWS services. With an instance role ARN set, any application that runs on an instance in this fleet can assume the role, including install scripts, server processes, and daemons (background processes). Create a role or look up a role's ARN by using the IAM dashboard in the AWS Management Console. Learn more about using on-box credentials for your game servers at Access external resources from a game server. This property cannot be changed after the fleet is created.

createFleet_metricGroups :: Lens' CreateFleet (Maybe [Text])

The name of an AWS CloudWatch metric group to add this fleet to. A metric group is used to aggregate the metrics for multiple fleets. You can specify an existing metric group name or set a new name to create a new metric group. A fleet can be included in only one metric group at a time.

createFleet_description :: Lens' CreateFleet (Maybe Text)

A human-readable description of the fleet.

createFleet_resourceCreationLimitPolicy :: Lens' CreateFleet (Maybe ResourceCreationLimitPolicy)

A policy that limits the number of game sessions that an individual player can create on instances in this fleet within a specified span of time.

createFleet_locations :: Lens' CreateFleet (Maybe (NonEmpty LocationConfiguration))

A set of remote locations to deploy additional instances to and manage as part of the fleet. This parameter can only be used when creating fleets in AWS Regions that support multiple locations. You can add any GameLift-supported AWS Region as a remote location, in the form of an AWS Region code such as us-west-2. To create a fleet with instances in the home Region only, omit this parameter.

createFleet_tags :: Lens' CreateFleet (Maybe [Tag])

A list of labels to assign to the new fleet resource. Tags are developer-defined key-value pairs. Tagging AWS resources are useful for resource management, access management and cost allocation. For more information, see Tagging AWS Resources in the AWS General Reference. Once the fleet is created, you can use TagResource, UntagResource, and ListTagsForResource to add, remove, and view tags. The maximum tag limit may be lower than stated. See the /AWS General Reference/ for actual tagging limits.

createFleet_name :: Lens' CreateFleet Text

A descriptive label that is associated with a fleet. Fleet names do not need to be unique.

createFleet_eC2InstanceType :: Lens' CreateFleet EC2InstanceType

The GameLift-supported EC2 instance type to use for all fleet instances. Instance type determines the computing resources that will be used to host your game servers, including CPU, memory, storage, and networking capacity. See Amazon EC2 Instance Types for detailed descriptions of EC2 instance types.

Destructuring the Response

data CreateFleetResponse

Represents the returned data in response to a request operation.

See: newCreateFleetResponse smart constructor.

Constructors

CreateFleetResponse'

Fields

locationStates :: Maybe [LocationState]

The fleet's locations and life-cycle status of each location. For new fleets, the status of all locations is set to NEW. During fleet creation, GameLift updates each location status as instances are deployed there and prepared for game hosting. This list includes an entry for the fleet's home Region. For fleets with no remote locations, only one entry, representing the home Region, is returned.

fleetAttributes :: Maybe FleetAttributes

The properties for the new fleet, including the current status. All fleets are placed in NEW status on creation.

httpStatus :: Int

The response's http status code.

newCreateFleetResponse

Arguments

:: Int
$sel:httpStatus:CreateFleetResponse'

-> CreateFleetResponse

Create a value of CreateFleetResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:locationStates:CreateFleetResponse', createFleetResponse_locationStates - The fleet's locations and life-cycle status of each location. For new fleets, the status of all locations is set to NEW. During fleet creation, GameLift updates each location status as instances are deployed there and prepared for game hosting. This list includes an entry for the fleet's home Region. For fleets with no remote locations, only one entry, representing the home Region, is returned.

$sel:fleetAttributes:CreateFleetResponse', createFleetResponse_fleetAttributes - The properties for the new fleet, including the current status. All fleets are placed in NEW status on creation.

$sel:httpStatus:CreateFleetResponse', createFleetResponse_httpStatus - The response's http status code.

Response Lenses

createFleetResponse_locationStates :: Lens' CreateFleetResponse (Maybe [LocationState])

The fleet's locations and life-cycle status of each location. For new fleets, the status of all locations is set to NEW. During fleet creation, GameLift updates each location status as instances are deployed there and prepared for game hosting. This list includes an entry for the fleet's home Region. For fleets with no remote locations, only one entry, representing the home Region, is returned.

createFleetResponse_fleetAttributes :: Lens' CreateFleetResponse (Maybe FleetAttributes)

The properties for the new fleet, including the current status. All fleets are placed in NEW status on creation.

createFleetResponse_httpStatus :: Lens' CreateFleetResponse Int

The response's http status code.