Amazonka.GameLift.CreateFleetLocations

Description

Adds remote locations to a fleet and begins populating the new locations with EC2 instances. The new instances conform to the fleet's instance type, auto-scaling, and other configuration settings.

This operation cannot be used with fleets that don't support remote locations. Fleets can have multiple locations only if they reside in AWS Regions that support this feature (see CreateFleet for the complete list) and were created after the feature was released in March 2021.

To add fleet locations, specify the fleet to be updated and provide a list of one or more locations.

If successful, this operation returns the list of added locations with their status set to NEW. GameLift initiates the process of starting an instance in each added location. You can track the status of each new location by monitoring location creation events using DescribeFleetEvents. Alternatively, you can poll location status by calling DescribeFleetLocationAttributes. After a location status becomes ACTIVE, you can adjust the location's capacity as needed with UpdateFleetCapacity.

Learn more

Setting up fleets

Multi-location fleets

Related actions

CreateFleetLocations | DescribeFleetLocationAttributes | DescribeFleetLocationCapacity | DescribeFleetLocationUtilization | DescribeFleetAttributes | DescribeFleetCapacity | DescribeFleetUtilization | UpdateFleetCapacity | StopFleetActions | DeleteFleetLocations | All APIs by task

Creating a Request

data CreateFleetLocations

Represents the input for a request operation.

See: newCreateFleetLocations smart constructor.

Constructors

CreateFleetLocations'

Fields

fleetId :: Text

A unique identifier for the fleet to add locations to. You can use either the fleet ID or ARN value.

locations :: NonEmpty LocationConfiguration

A list of locations to deploy additional instances to and manage as part of the fleet. You can add any GameLift-supported AWS Region as a remote location, in the form of an AWS Region code such as us-west-2.

newCreateFleetLocations

Arguments

:: Text
$sel:fleetId:CreateFleetLocations'

-> NonEmpty LocationConfiguration
$sel:locations:CreateFleetLocations'

-> CreateFleetLocations

Create a value of CreateFleetLocations with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:fleetId:CreateFleetLocations', createFleetLocations_fleetId - A unique identifier for the fleet to add locations to. You can use either the fleet ID or ARN value.

$sel:locations:CreateFleetLocations', createFleetLocations_locations - A list of locations to deploy additional instances to and manage as part of the fleet. You can add any GameLift-supported AWS Region as a remote location, in the form of an AWS Region code such as us-west-2.

Request Lenses

createFleetLocations_fleetId :: Lens' CreateFleetLocations Text

A unique identifier for the fleet to add locations to. You can use either the fleet ID or ARN value.

createFleetLocations_locations :: Lens' CreateFleetLocations (NonEmpty LocationConfiguration)

A list of locations to deploy additional instances to and manage as part of the fleet. You can add any GameLift-supported AWS Region as a remote location, in the form of an AWS Region code such as us-west-2.

Destructuring the Response

data CreateFleetLocationsResponse

Represents the returned data in response to a request operation.

See: newCreateFleetLocationsResponse smart constructor.

Constructors

CreateFleetLocationsResponse'

Fields

fleetArn :: Maybe Text

The Amazon Resource Name (ARN) that is assigned to a GameLift fleet resource and uniquely identifies it. ARNs are unique across all Regions. Format is arn:aws:gamelift:<region>::fleet/fleet-a1234567-b8c9-0d1e-2fa3-b45c6d7e8912.

locationStates :: Maybe [LocationState]

The remote locations that are being added to the fleet, and the life-cycle status of each location. For new locations, the status is set to NEW. During location creation, GameLift updates each location's status as instances are deployed there and prepared for game hosting. This list does not include the fleet home Region or any remote locations that were already added to the fleet.

fleetId :: Maybe Text

A unique identifier for the fleet that was updated with new locations.

httpStatus :: Int

The response's http status code.

newCreateFleetLocationsResponse

Arguments

:: Int
$sel:httpStatus:CreateFleetLocationsResponse'

-> CreateFleetLocationsResponse

Create a value of CreateFleetLocationsResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:fleetArn:CreateFleetLocationsResponse', createFleetLocationsResponse_fleetArn - The Amazon Resource Name (ARN) that is assigned to a GameLift fleet resource and uniquely identifies it. ARNs are unique across all Regions. Format is arn:aws:gamelift:<region>::fleet/fleet-a1234567-b8c9-0d1e-2fa3-b45c6d7e8912.

$sel:locationStates:CreateFleetLocationsResponse', createFleetLocationsResponse_locationStates - The remote locations that are being added to the fleet, and the life-cycle status of each location. For new locations, the status is set to NEW. During location creation, GameLift updates each location's status as instances are deployed there and prepared for game hosting. This list does not include the fleet home Region or any remote locations that were already added to the fleet.

$sel:fleetId:CreateFleetLocations', createFleetLocationsResponse_fleetId - A unique identifier for the fleet that was updated with new locations.

$sel:httpStatus:CreateFleetLocationsResponse', createFleetLocationsResponse_httpStatus - The response's http status code.

Response Lenses

createFleetLocationsResponse_fleetArn :: Lens' CreateFleetLocationsResponse (Maybe Text)

The Amazon Resource Name (ARN) that is assigned to a GameLift fleet resource and uniquely identifies it. ARNs are unique across all Regions. Format is arn:aws:gamelift:<region>::fleet/fleet-a1234567-b8c9-0d1e-2fa3-b45c6d7e8912.

createFleetLocationsResponse_locationStates :: Lens' CreateFleetLocationsResponse (Maybe [LocationState])

The remote locations that are being added to the fleet, and the life-cycle status of each location. For new locations, the status is set to NEW. During location creation, GameLift updates each location's status as instances are deployed there and prepared for game hosting. This list does not include the fleet home Region or any remote locations that were already added to the fleet.

createFleetLocationsResponse_fleetId :: Lens' CreateFleetLocationsResponse (Maybe Text)

A unique identifier for the fleet that was updated with new locations.

createFleetLocationsResponse_httpStatus :: Lens' CreateFleetLocationsResponse Int

The response's http status code.