Amazonka.GameLift.DeregisterGameServer

Description

__This operation is used with the GameLift FleetIQ solution and game server groups.__

Removes the game server from a game server group. As a result of this operation, the deregistered game server can no longer be claimed and will not be returned in a list of active game servers.

To deregister a game server, specify the game server group and game server ID. If successful, this operation emits a CloudWatch event with termination timestamp and reason.

Learn more

GameLift FleetIQ Guide

Related actions

RegisterGameServer | ListGameServers | ClaimGameServer | DescribeGameServer | UpdateGameServer | DeregisterGameServer | All APIs by task

Creating a Request

data DeregisterGameServer

See: newDeregisterGameServer smart constructor.

Constructors

DeregisterGameServer'

Fields

gameServerGroupName :: Text

A unique identifier for the game server group where the game server is running. Use either the GameServerGroup name or ARN value.

gameServerId :: Text

A custom string that uniquely identifies the game server to deregister.

newDeregisterGameServer

Arguments

:: Text
$sel:gameServerGroupName:DeregisterGameServer'

-> Text
$sel:gameServerId:DeregisterGameServer'

-> DeregisterGameServer

Create a value of DeregisterGameServer with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:gameServerGroupName:DeregisterGameServer', deregisterGameServer_gameServerGroupName - A unique identifier for the game server group where the game server is running. Use either the GameServerGroup name or ARN value.

$sel:gameServerId:DeregisterGameServer', deregisterGameServer_gameServerId - A custom string that uniquely identifies the game server to deregister.

Request Lenses

deregisterGameServer_gameServerGroupName :: Lens' DeregisterGameServer Text

A unique identifier for the game server group where the game server is running. Use either the GameServerGroup name or ARN value.

deregisterGameServer_gameServerId :: Lens' DeregisterGameServer Text

A custom string that uniquely identifies the game server to deregister.

Destructuring the Response

data DeregisterGameServerResponse

See: newDeregisterGameServerResponse smart constructor.

Constructors

DeregisterGameServerResponse'

newDeregisterGameServerResponse :: DeregisterGameServerResponse

Create a value of DeregisterGameServerResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.