Amazonka.GameLift.DescribeFleetLocationCapacity

Description

Retrieves the resource capacity settings for a fleet location. The data returned includes the current capacity (number of EC2 instances) and some scaling settings for the requested fleet location. Use this operation to retrieve capacity information for a fleet's remote location or home Region (you can also retrieve home Region capacity by calling DescribeFleetCapacity).

To retrieve capacity data, identify a fleet and location.

If successful, a FleetCapacity object is returned for the requested fleet location.

Learn more

Setting up GameLift fleets

GameLift metrics for fleets

Related actions

CreateFleetLocations | DescribeFleetLocationAttributes | DescribeFleetLocationCapacity | DescribeFleetLocationUtilization | DescribeFleetAttributes | DescribeFleetCapacity | DescribeFleetUtilization | UpdateFleetCapacity | StopFleetActions | DeleteFleetLocations | All APIs by task

Creating a Request

data DescribeFleetLocationCapacity

Represents the input for a request operation.

See: newDescribeFleetLocationCapacity smart constructor.

Constructors

DescribeFleetLocationCapacity'

Fields

fleetId :: Text

A unique identifier for the fleet to request location capacity for. You can use either the fleet ID or ARN value.

location :: Text

The fleet location to retrieve capacity information for. Specify a location in the form of an AWS Region code, such as us-west-2.

newDescribeFleetLocationCapacity

Arguments

:: Text
$sel:fleetId:DescribeFleetLocationCapacity'

-> Text
$sel:location:DescribeFleetLocationCapacity'

-> DescribeFleetLocationCapacity

Create a value of DescribeFleetLocationCapacity with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:fleetId:DescribeFleetLocationCapacity', describeFleetLocationCapacity_fleetId - A unique identifier for the fleet to request location capacity for. You can use either the fleet ID or ARN value.

$sel:location:DescribeFleetLocationCapacity', describeFleetLocationCapacity_location - The fleet location to retrieve capacity information for. Specify a location in the form of an AWS Region code, such as us-west-2.

Request Lenses

describeFleetLocationCapacity_fleetId :: Lens' DescribeFleetLocationCapacity Text

A unique identifier for the fleet to request location capacity for. You can use either the fleet ID or ARN value.

describeFleetLocationCapacity_location :: Lens' DescribeFleetLocationCapacity Text

The fleet location to retrieve capacity information for. Specify a location in the form of an AWS Region code, such as us-west-2.

Destructuring the Response

data DescribeFleetLocationCapacityResponse

Represents the returned data in response to a request operation.

See: newDescribeFleetLocationCapacityResponse smart constructor.

Constructors

DescribeFleetLocationCapacityResponse'

Fields

fleetCapacity :: Maybe FleetCapacity

Resource capacity information for the requested fleet location. Capacity objects are returned only for fleets and locations that currently exist.

httpStatus :: Int

The response's http status code.

newDescribeFleetLocationCapacityResponse

Arguments

:: Int
$sel:httpStatus:DescribeFleetLocationCapacityResponse'

-> DescribeFleetLocationCapacityResponse

Create a value of DescribeFleetLocationCapacityResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:fleetCapacity:DescribeFleetLocationCapacityResponse', describeFleetLocationCapacityResponse_fleetCapacity - Resource capacity information for the requested fleet location. Capacity objects are returned only for fleets and locations that currently exist.

$sel:httpStatus:DescribeFleetLocationCapacityResponse', describeFleetLocationCapacityResponse_httpStatus - The response's http status code.

Response Lenses

describeFleetLocationCapacityResponse_fleetCapacity :: Lens' DescribeFleetLocationCapacityResponse (Maybe FleetCapacity)

Resource capacity information for the requested fleet location. Capacity objects are returned only for fleets and locations that currently exist.

describeFleetLocationCapacityResponse_httpStatus :: Lens' DescribeFleetLocationCapacityResponse Int

The response's http status code.