Amazonka.GameLift.DescribeFleetPortSettings

Description

Retrieves a fleet's inbound connection permissions. Connection permissions specify the range of IP addresses and port settings that incoming traffic can use to access server processes in the fleet. Game sessions that are running on instances in the fleet must use connections that fall in this range.

This operation can be used in the following ways:

To retrieve the inbound connection permissions for a fleet, identify the fleet's unique identifier.

To check the status of recent updates to a fleet remote location, specify the fleet ID and a location. Port setting updates can take time to propagate across all locations.

If successful, a set of IpPermission objects is returned for the requested fleet ID. When a location is specified, a pending status is included. If the requested fleet has been deleted, the result set is empty.

Learn more

Setting up GameLift fleets

Related actions

ListFleets | DescribeEC2InstanceLimits | DescribeFleetAttributes | DescribeFleetCapacity | DescribeFleetEvents | DescribeFleetLocationAttributes | DescribeFleetPortSettings | DescribeFleetUtilization | DescribeRuntimeConfiguration | DescribeScalingPolicies | All APIs by task

Creating a Request

data DescribeFleetPortSettings

Represents the input for a request operation.

See: newDescribeFleetPortSettings smart constructor.

Constructors

DescribeFleetPortSettings'

Fields

location :: Maybe Text

A remote location to check for status of port setting updates. Use the AWS Region code format, such as us-west-2.

fleetId :: Text

A unique identifier for the fleet to retrieve port settings for. You can use either the fleet ID or ARN value.

newDescribeFleetPortSettings

Arguments

:: Text
$sel:fleetId:DescribeFleetPortSettings'

-> DescribeFleetPortSettings

Create a value of DescribeFleetPortSettings with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:location:DescribeFleetPortSettings', describeFleetPortSettings_location - A remote location to check for status of port setting updates. Use the AWS Region code format, such as us-west-2.

$sel:fleetId:DescribeFleetPortSettings', describeFleetPortSettings_fleetId - A unique identifier for the fleet to retrieve port settings for. You can use either the fleet ID or ARN value.

Request Lenses

describeFleetPortSettings_location :: Lens' DescribeFleetPortSettings (Maybe Text)

A remote location to check for status of port setting updates. Use the AWS Region code format, such as us-west-2.

describeFleetPortSettings_fleetId :: Lens' DescribeFleetPortSettings Text

A unique identifier for the fleet to retrieve port settings for. You can use either the fleet ID or ARN value.

Destructuring the Response

data DescribeFleetPortSettingsResponse

Represents the returned data in response to a request operation.

See: newDescribeFleetPortSettingsResponse smart constructor.

Constructors

DescribeFleetPortSettingsResponse'

Fields

location :: Maybe Text

The requested fleet location, expressed as an AWS Region code, such as us-west-2.

inboundPermissions :: Maybe [IpPermission]

The port settings for the requested fleet ID.

fleetArn :: Maybe Text

The Amazon Resource Name (ARN) that is assigned to a GameLift fleet resource and uniquely identifies it. ARNs are unique across all Regions. Format is arn:aws:gamelift:<region>::fleet/fleet-a1234567-b8c9-0d1e-2fa3-b45c6d7e8912.

updateStatus :: Maybe LocationUpdateStatus

The current status of updates to the fleet's port settings in the requested fleet location. A status of PENDING_UPDATE indicates that an update was requested for the fleet but has not yet been completed for the location.

fleetId :: Maybe Text

A unique identifier for the fleet that was requested.

httpStatus :: Int

The response's http status code.

newDescribeFleetPortSettingsResponse

Arguments

:: Int
$sel:httpStatus:DescribeFleetPortSettingsResponse'

-> DescribeFleetPortSettingsResponse

Create a value of DescribeFleetPortSettingsResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:location:DescribeFleetPortSettings', describeFleetPortSettingsResponse_location - The requested fleet location, expressed as an AWS Region code, such as us-west-2.

$sel:inboundPermissions:DescribeFleetPortSettingsResponse', describeFleetPortSettingsResponse_inboundPermissions - The port settings for the requested fleet ID.

$sel:fleetArn:DescribeFleetPortSettingsResponse', describeFleetPortSettingsResponse_fleetArn - The Amazon Resource Name (ARN) that is assigned to a GameLift fleet resource and uniquely identifies it. ARNs are unique across all Regions. Format is arn:aws:gamelift:<region>::fleet/fleet-a1234567-b8c9-0d1e-2fa3-b45c6d7e8912.

$sel:updateStatus:DescribeFleetPortSettingsResponse', describeFleetPortSettingsResponse_updateStatus - The current status of updates to the fleet's port settings in the requested fleet location. A status of PENDING_UPDATE indicates that an update was requested for the fleet but has not yet been completed for the location.

$sel:fleetId:DescribeFleetPortSettings', describeFleetPortSettingsResponse_fleetId - A unique identifier for the fleet that was requested.

$sel:httpStatus:DescribeFleetPortSettingsResponse', describeFleetPortSettingsResponse_httpStatus - The response's http status code.

Response Lenses

describeFleetPortSettingsResponse_location :: Lens' DescribeFleetPortSettingsResponse (Maybe Text)

The requested fleet location, expressed as an AWS Region code, such as us-west-2.

describeFleetPortSettingsResponse_inboundPermissions :: Lens' DescribeFleetPortSettingsResponse (Maybe [IpPermission])

The port settings for the requested fleet ID.

describeFleetPortSettingsResponse_fleetArn :: Lens' DescribeFleetPortSettingsResponse (Maybe Text)

The Amazon Resource Name (ARN) that is assigned to a GameLift fleet resource and uniquely identifies it. ARNs are unique across all Regions. Format is arn:aws:gamelift:<region>::fleet/fleet-a1234567-b8c9-0d1e-2fa3-b45c6d7e8912.

describeFleetPortSettingsResponse_updateStatus :: Lens' DescribeFleetPortSettingsResponse (Maybe LocationUpdateStatus)

The current status of updates to the fleet's port settings in the requested fleet location. A status of PENDING_UPDATE indicates that an update was requested for the fleet but has not yet been completed for the location.

describeFleetPortSettingsResponse_fleetId :: Lens' DescribeFleetPortSettingsResponse (Maybe Text)

A unique identifier for the fleet that was requested.

describeFleetPortSettingsResponse_httpStatus :: Lens' DescribeFleetPortSettingsResponse Int

The response's http status code.