Amazonka.GameLift.ListBuilds

Description

Retrieves build resources for all builds associated with the AWS account in use. You can limit results to builds that are in a specific status by using the Status parameter. Use the pagination parameters to retrieve results in a set of sequential pages.

Build resources are not listed in any particular order.

Learn more

Upload a Custom Server Build

Related actions

CreateBuild | ListBuilds | DescribeBuild | UpdateBuild | DeleteBuild | All APIs by task

This operation returns paginated results.

Creating a Request

data ListBuilds

Represents the input for a request operation.

See: newListBuilds smart constructor.

Constructors

ListBuilds'

Fields

status :: Maybe BuildStatus

Build status to filter results by. To retrieve all builds, leave this parameter empty.

Possible build statuses include the following:

INITIALIZED -- A new build has been defined, but no files have been uploaded. You cannot create fleets for builds that are in this status. When a build is successfully created, the build status is set to this value.

READY -- The game build has been successfully uploaded. You can now create new fleets for this build.

FAILED -- The game build upload failed. You cannot create new fleets for this build.

nextToken :: Maybe Text

A token that indicates the start of the next sequential page of results. Use the token that is returned with a previous call to this operation. To start at the beginning of the result set, do not specify a value.

limit :: Maybe Natural

The maximum number of results to return. Use this parameter with NextToken to get results as a set of sequential pages.

newListBuilds :: ListBuilds

Create a value of ListBuilds with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:status:ListBuilds', listBuilds_status - Build status to filter results by. To retrieve all builds, leave this parameter empty.

Possible build statuses include the following:

INITIALIZED -- A new build has been defined, but no files have been uploaded. You cannot create fleets for builds that are in this status. When a build is successfully created, the build status is set to this value.

READY -- The game build has been successfully uploaded. You can now create new fleets for this build.

FAILED -- The game build upload failed. You cannot create new fleets for this build.

$sel:nextToken:ListBuilds', listBuilds_nextToken - A token that indicates the start of the next sequential page of results. Use the token that is returned with a previous call to this operation. To start at the beginning of the result set, do not specify a value.

$sel:limit:ListBuilds', listBuilds_limit - The maximum number of results to return. Use this parameter with NextToken to get results as a set of sequential pages.

Request Lenses

listBuilds_status :: Lens' ListBuilds (Maybe BuildStatus)

Build status to filter results by. To retrieve all builds, leave this parameter empty.

Possible build statuses include the following:

INITIALIZED -- A new build has been defined, but no files have been uploaded. You cannot create fleets for builds that are in this status. When a build is successfully created, the build status is set to this value.

READY -- The game build has been successfully uploaded. You can now create new fleets for this build.

FAILED -- The game build upload failed. You cannot create new fleets for this build.

listBuilds_nextToken :: Lens' ListBuilds (Maybe Text)

A token that indicates the start of the next sequential page of results. Use the token that is returned with a previous call to this operation. To start at the beginning of the result set, do not specify a value.

listBuilds_limit :: Lens' ListBuilds (Maybe Natural)

The maximum number of results to return. Use this parameter with NextToken to get results as a set of sequential pages.

Destructuring the Response

data ListBuildsResponse

Represents the returned data in response to a request operation.

See: newListBuildsResponse smart constructor.

Constructors

ListBuildsResponse'

Fields

builds :: Maybe [Build]

A collection of build resources that match the request.

nextToken :: Maybe Text

A token that indicates where to resume retrieving results on the next call to this operation. If no token is returned, these results represent the end of the list.

httpStatus :: Int

The response's http status code.

newListBuildsResponse

Arguments

:: Int
$sel:httpStatus:ListBuildsResponse'

-> ListBuildsResponse

Create a value of ListBuildsResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:builds:ListBuildsResponse', listBuildsResponse_builds - A collection of build resources that match the request.

$sel:nextToken:ListBuilds', listBuildsResponse_nextToken - A token that indicates where to resume retrieving results on the next call to this operation. If no token is returned, these results represent the end of the list.

$sel:httpStatus:ListBuildsResponse', listBuildsResponse_httpStatus - The response's http status code.

Response Lenses

listBuildsResponse_builds :: Lens' ListBuildsResponse (Maybe [Build])

A collection of build resources that match the request.

listBuildsResponse_nextToken :: Lens' ListBuildsResponse (Maybe Text)

A token that indicates where to resume retrieving results on the next call to this operation. If no token is returned, these results represent the end of the list.

listBuildsResponse_httpStatus :: Lens' ListBuildsResponse Int

The response's http status code.