Amazonka.GameLift.ListFleets

Description

Retrieves a collection of fleet resources in an AWS Region. You can call this operation to get fleets in a previously selected default Region (see https://docs.aws.amazon.com/credref/latest/refdocs/setting-global-region.htmlor specify a Region in your request. You can filter the result set to find only those fleets that are deployed with a specific build or script. For fleets that have multiple locations, this operation retrieves fleets based on their home Region only.

This operation can be used in the following ways:

To get a list of all fleets in a Region, don't provide a build or script identifier.

To get a list of all fleets where a specific custom game build is deployed, provide the build ID.

To get a list of all Realtime Servers fleets with a specific configuration script, provide the script ID.

Use the pagination parameters to retrieve results as a set of sequential pages.

If successful, a list of fleet IDs that match the request parameters is returned. A NextToken value is also returned if there are more result pages to retrieve.

Fleet resources are not listed in a particular order.

Learn more

Setting up GameLift fleets

Related actions

CreateFleet | UpdateFleetCapacity | PutScalingPolicy | DescribeEC2InstanceLimits | DescribeFleetAttributes | DescribeFleetLocationAttributes | UpdateFleetAttributes | StopFleetActions | DeleteFleet | All APIs by task

This operation returns paginated results.

Creating a Request

data ListFleets

Represents the input for a request operation.

See: newListFleets smart constructor.

Constructors

ListFleets'

Fields

buildId :: Maybe Text

A unique identifier for the build to request fleets for. Use this parameter to return only fleets using a specified build. Use either the build ID or ARN value.

nextToken :: Maybe Text

A token that indicates the start of the next sequential page of results. Use the token that is returned with a previous call to this operation. To start at the beginning of the result set, do not specify a value.

scriptId :: Maybe Text

A unique identifier for the Realtime script to request fleets for. Use this parameter to return only fleets using a specified script. Use either the script ID or ARN value.

limit :: Maybe Natural

The maximum number of results to return. Use this parameter with NextToken to get results as a set of sequential pages.

newListFleets :: ListFleets

Create a value of ListFleets with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:buildId:ListFleets', listFleets_buildId - A unique identifier for the build to request fleets for. Use this parameter to return only fleets using a specified build. Use either the build ID or ARN value.

$sel:nextToken:ListFleets', listFleets_nextToken - A token that indicates the start of the next sequential page of results. Use the token that is returned with a previous call to this operation. To start at the beginning of the result set, do not specify a value.

$sel:scriptId:ListFleets', listFleets_scriptId - A unique identifier for the Realtime script to request fleets for. Use this parameter to return only fleets using a specified script. Use either the script ID or ARN value.

$sel:limit:ListFleets', listFleets_limit - The maximum number of results to return. Use this parameter with NextToken to get results as a set of sequential pages.

Request Lenses

listFleets_buildId :: Lens' ListFleets (Maybe Text)

A unique identifier for the build to request fleets for. Use this parameter to return only fleets using a specified build. Use either the build ID or ARN value.

listFleets_nextToken :: Lens' ListFleets (Maybe Text)

A token that indicates the start of the next sequential page of results. Use the token that is returned with a previous call to this operation. To start at the beginning of the result set, do not specify a value.

listFleets_scriptId :: Lens' ListFleets (Maybe Text)

A unique identifier for the Realtime script to request fleets for. Use this parameter to return only fleets using a specified script. Use either the script ID or ARN value.

listFleets_limit :: Lens' ListFleets (Maybe Natural)

The maximum number of results to return. Use this parameter with NextToken to get results as a set of sequential pages.

Destructuring the Response

data ListFleetsResponse

Represents the returned data in response to a request operation.

See: newListFleetsResponse smart constructor.

Constructors

ListFleetsResponse'

Fields

nextToken :: Maybe Text

A token that indicates where to resume retrieving results on the next call to this operation. If no token is returned, these results represent the end of the list.

fleetIds :: Maybe (NonEmpty Text)

A set of fleet IDs that match the list request. You can retrieve additional information about all returned fleets by passing this result set to a DescribeFleetAttributes, DescribeFleetCapacity, or DescribeFleetUtilization call.

httpStatus :: Int

The response's http status code.

newListFleetsResponse

Arguments

:: Int
$sel:httpStatus:ListFleetsResponse'

-> ListFleetsResponse

Create a value of ListFleetsResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:nextToken:ListFleets', listFleetsResponse_nextToken - A token that indicates where to resume retrieving results on the next call to this operation. If no token is returned, these results represent the end of the list.

$sel:fleetIds:ListFleetsResponse', listFleetsResponse_fleetIds - A set of fleet IDs that match the list request. You can retrieve additional information about all returned fleets by passing this result set to a DescribeFleetAttributes, DescribeFleetCapacity, or DescribeFleetUtilization call.

$sel:httpStatus:ListFleetsResponse', listFleetsResponse_httpStatus - The response's http status code.

Response Lenses

listFleetsResponse_nextToken :: Lens' ListFleetsResponse (Maybe Text)

A token that indicates where to resume retrieving results on the next call to this operation. If no token is returned, these results represent the end of the list.

listFleetsResponse_fleetIds :: Lens' ListFleetsResponse (Maybe (NonEmpty Text))

A set of fleet IDs that match the list request. You can retrieve additional information about all returned fleets by passing this result set to a DescribeFleetAttributes, DescribeFleetCapacity, or DescribeFleetUtilization call.

listFleetsResponse_httpStatus :: Lens' ListFleetsResponse Int

The response's http status code.