Amazonka.GameLift.Types.GameSessionConnectionInfo

Description

Documentation

data GameSessionConnectionInfo

Connection information for a new game session that is created in response to a StartMatchmaking request. Once a match is made, the FlexMatch engine creates a new game session for it. This information, including the game session endpoint and player sessions for each player in the original matchmaking request, is added to the MatchmakingTicket, which can be retrieved by calling DescribeMatchmaking.

See: newGameSessionConnectionInfo smart constructor.

Constructors

GameSessionConnectionInfo'

Fields

matchedPlayerSessions :: Maybe [MatchedPlayerSession]

A collection of player session IDs, one for each player ID that was included in the original matchmaking request.

ipAddress :: Maybe Text

The IP address of the game session. To connect to a GameLift game server, an app needs both the IP address and port number.

gameSessionArn :: Maybe Text

A unique identifier for the game session. Use the game session ID.

dnsName :: Maybe Text

The DNS identifier assigned to the instance that is running the game session. Values have the following format:

TLS-enabled fleets: <unique identifier>.<region identifier>.amazongamelift.com.

Non-TLS-enabled fleets: ec2-<unique identifier>.compute.amazonaws.com. (See Amazon EC2 Instance IP Addressing.)

When connecting to a game session that is running on a TLS-enabled fleet, you must use the DNS name, not the IP address.

port :: Maybe Natural

The port number for the game session. To connect to a GameLift game server, an app needs both the IP address and port number.

newGameSessionConnectionInfo :: GameSessionConnectionInfo

Create a value of GameSessionConnectionInfo with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:matchedPlayerSessions:GameSessionConnectionInfo', gameSessionConnectionInfo_matchedPlayerSessions - A collection of player session IDs, one for each player ID that was included in the original matchmaking request.

$sel:ipAddress:GameSessionConnectionInfo', gameSessionConnectionInfo_ipAddress - The IP address of the game session. To connect to a GameLift game server, an app needs both the IP address and port number.

$sel:gameSessionArn:GameSessionConnectionInfo', gameSessionConnectionInfo_gameSessionArn - A unique identifier for the game session. Use the game session ID.

$sel:dnsName:GameSessionConnectionInfo', gameSessionConnectionInfo_dnsName - The DNS identifier assigned to the instance that is running the game session. Values have the following format:

TLS-enabled fleets: <unique identifier>.<region identifier>.amazongamelift.com.

Non-TLS-enabled fleets: ec2-<unique identifier>.compute.amazonaws.com. (See Amazon EC2 Instance IP Addressing.)

When connecting to a game session that is running on a TLS-enabled fleet, you must use the DNS name, not the IP address.

$sel:port:GameSessionConnectionInfo', gameSessionConnectionInfo_port - The port number for the game session. To connect to a GameLift game server, an app needs both the IP address and port number.

gameSessionConnectionInfo_matchedPlayerSessions :: Lens' GameSessionConnectionInfo (Maybe [MatchedPlayerSession])

A collection of player session IDs, one for each player ID that was included in the original matchmaking request.

gameSessionConnectionInfo_ipAddress :: Lens' GameSessionConnectionInfo (Maybe Text)

The IP address of the game session. To connect to a GameLift game server, an app needs both the IP address and port number.

gameSessionConnectionInfo_gameSessionArn :: Lens' GameSessionConnectionInfo (Maybe Text)

A unique identifier for the game session. Use the game session ID.

gameSessionConnectionInfo_dnsName :: Lens' GameSessionConnectionInfo (Maybe Text)

The DNS identifier assigned to the instance that is running the game session. Values have the following format:

TLS-enabled fleets: <unique identifier>.<region identifier>.amazongamelift.com.

Non-TLS-enabled fleets: ec2-<unique identifier>.compute.amazonaws.com. (See Amazon EC2 Instance IP Addressing.)

When connecting to a game session that is running on a TLS-enabled fleet, you must use the DNS name, not the IP address.

gameSessionConnectionInfo_port :: Lens' GameSessionConnectionInfo (Maybe Natural)

The port number for the game session. To connect to a GameLift game server, an app needs both the IP address and port number.