Amazonka.GameLift.Types.MatchmakingTicket

Description

Documentation

data MatchmakingTicket

Ticket generated to track the progress of a matchmaking request. Each ticket is uniquely identified by a ticket ID, supplied by the requester, when creating a matchmaking request with StartMatchmaking. Tickets can be retrieved by calling DescribeMatchmaking with the ticket ID.

See: newMatchmakingTicket smart constructor.

Constructors

MatchmakingTicket'

Fields

status :: Maybe MatchmakingConfigurationStatus

Current status of the matchmaking request.

QUEUED -- The matchmaking request has been received and is currently waiting to be processed.

SEARCHING -- The matchmaking request is currently being processed.

REQUIRES_ACCEPTANCE -- A match has been proposed and the players must accept the match (see AcceptMatch). This status is used only with requests that use a matchmaking configuration with a player acceptance requirement.

PLACING -- The FlexMatch engine has matched players and is in the process of placing a new game session for the match.

COMPLETED -- Players have been matched and a game session is ready to host the players. A ticket in this state contains the necessary connection information for players.

FAILED -- The matchmaking request was not completed.

CANCELLED -- The matchmaking request was canceled. This may be the result of a call to StopMatchmaking or a proposed match that one or more players failed to accept.

TIMED_OUT -- The matchmaking request was not successful within the duration specified in the matchmaking configuration.

Matchmaking requests that fail to successfully complete (statuses FAILED, CANCELLED, TIMED_OUT) can be resubmitted as new requests with new ticket IDs.

configurationName :: Maybe Text

Name of the MatchmakingConfiguration that is used with this ticket. Matchmaking configurations determine how players are grouped into a match and how a new game session is created for the match.

startTime :: Maybe POSIX

Time stamp indicating when this matchmaking request was received. Format is a number expressed in Unix time as milliseconds (for example "1469498468.057").

gameSessionConnectionInfo :: Maybe GameSessionConnectionInfo

Identifier and connection information of the game session created for the match. This information is added to the ticket only after the matchmaking request has been successfully completed. This parameter is not set when FlexMatch is being used without GameLift hosting.

ticketId :: Maybe Text

A unique identifier for a matchmaking ticket.

estimatedWaitTime :: Maybe Natural

Average amount of time (in seconds) that players are currently waiting for a match. If there is not enough recent data, this property may be empty.

statusMessage :: Maybe Text

Additional information about the current status.

endTime :: Maybe POSIX

Time stamp indicating when this matchmaking request stopped being processed due to success, failure, or cancellation. Format is a number expressed in Unix time as milliseconds (for example "1469498468.057").

configurationArn :: Maybe Text

The Amazon Resource Name (ARN) associated with the GameLift matchmaking configuration resource that is used with this ticket.

statusReason :: Maybe Text

Code to explain the current status. For example, a status reason may indicate when a ticket has returned to SEARCHING status after a proposed match fails to receive player acceptances.

players :: Maybe [Player]

A set of Player objects, each representing a player to find matches for. Players are identified by a unique player ID and may include latency data for use during matchmaking. If the ticket is in status COMPLETED, the Player objects include the team the players were assigned to in the resulting match.

newMatchmakingTicket :: MatchmakingTicket

Create a value of MatchmakingTicket with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:status:MatchmakingTicket', matchmakingTicket_status - Current status of the matchmaking request.

QUEUED -- The matchmaking request has been received and is currently waiting to be processed.

SEARCHING -- The matchmaking request is currently being processed.

REQUIRES_ACCEPTANCE -- A match has been proposed and the players must accept the match (see AcceptMatch). This status is used only with requests that use a matchmaking configuration with a player acceptance requirement.

PLACING -- The FlexMatch engine has matched players and is in the process of placing a new game session for the match.

COMPLETED -- Players have been matched and a game session is ready to host the players. A ticket in this state contains the necessary connection information for players.

FAILED -- The matchmaking request was not completed.

CANCELLED -- The matchmaking request was canceled. This may be the result of a call to StopMatchmaking or a proposed match that one or more players failed to accept.

TIMED_OUT -- The matchmaking request was not successful within the duration specified in the matchmaking configuration.

Matchmaking requests that fail to successfully complete (statuses FAILED, CANCELLED, TIMED_OUT) can be resubmitted as new requests with new ticket IDs.

$sel:configurationName:MatchmakingTicket', matchmakingTicket_configurationName - Name of the MatchmakingConfiguration that is used with this ticket. Matchmaking configurations determine how players are grouped into a match and how a new game session is created for the match.

$sel:startTime:MatchmakingTicket', matchmakingTicket_startTime - Time stamp indicating when this matchmaking request was received. Format is a number expressed in Unix time as milliseconds (for example "1469498468.057").

$sel:gameSessionConnectionInfo:MatchmakingTicket', matchmakingTicket_gameSessionConnectionInfo - Identifier and connection information of the game session created for the match. This information is added to the ticket only after the matchmaking request has been successfully completed. This parameter is not set when FlexMatch is being used without GameLift hosting.

$sel:ticketId:MatchmakingTicket', matchmakingTicket_ticketId - A unique identifier for a matchmaking ticket.

$sel:estimatedWaitTime:MatchmakingTicket', matchmakingTicket_estimatedWaitTime - Average amount of time (in seconds) that players are currently waiting for a match. If there is not enough recent data, this property may be empty.

$sel:statusMessage:MatchmakingTicket', matchmakingTicket_statusMessage - Additional information about the current status.

$sel:endTime:MatchmakingTicket', matchmakingTicket_endTime - Time stamp indicating when this matchmaking request stopped being processed due to success, failure, or cancellation. Format is a number expressed in Unix time as milliseconds (for example "1469498468.057").

$sel:configurationArn:MatchmakingTicket', matchmakingTicket_configurationArn - The Amazon Resource Name (ARN) associated with the GameLift matchmaking configuration resource that is used with this ticket.

$sel:statusReason:MatchmakingTicket', matchmakingTicket_statusReason - Code to explain the current status. For example, a status reason may indicate when a ticket has returned to SEARCHING status after a proposed match fails to receive player acceptances.

$sel:players:MatchmakingTicket', matchmakingTicket_players - A set of Player objects, each representing a player to find matches for. Players are identified by a unique player ID and may include latency data for use during matchmaking. If the ticket is in status COMPLETED, the Player objects include the team the players were assigned to in the resulting match.

matchmakingTicket_status :: Lens' MatchmakingTicket (Maybe MatchmakingConfigurationStatus)

Current status of the matchmaking request.

QUEUED -- The matchmaking request has been received and is currently waiting to be processed.

SEARCHING -- The matchmaking request is currently being processed.

REQUIRES_ACCEPTANCE -- A match has been proposed and the players must accept the match (see AcceptMatch). This status is used only with requests that use a matchmaking configuration with a player acceptance requirement.

PLACING -- The FlexMatch engine has matched players and is in the process of placing a new game session for the match.

COMPLETED -- Players have been matched and a game session is ready to host the players. A ticket in this state contains the necessary connection information for players.

FAILED -- The matchmaking request was not completed.

CANCELLED -- The matchmaking request was canceled. This may be the result of a call to StopMatchmaking or a proposed match that one or more players failed to accept.

TIMED_OUT -- The matchmaking request was not successful within the duration specified in the matchmaking configuration.

Matchmaking requests that fail to successfully complete (statuses FAILED, CANCELLED, TIMED_OUT) can be resubmitted as new requests with new ticket IDs.

matchmakingTicket_configurationName :: Lens' MatchmakingTicket (Maybe Text)

Name of the MatchmakingConfiguration that is used with this ticket. Matchmaking configurations determine how players are grouped into a match and how a new game session is created for the match.

matchmakingTicket_startTime :: Lens' MatchmakingTicket (Maybe UTCTime)

Time stamp indicating when this matchmaking request was received. Format is a number expressed in Unix time as milliseconds (for example "1469498468.057").

matchmakingTicket_gameSessionConnectionInfo :: Lens' MatchmakingTicket (Maybe GameSessionConnectionInfo)

Identifier and connection information of the game session created for the match. This information is added to the ticket only after the matchmaking request has been successfully completed. This parameter is not set when FlexMatch is being used without GameLift hosting.

matchmakingTicket_ticketId :: Lens' MatchmakingTicket (Maybe Text)

A unique identifier for a matchmaking ticket.

matchmakingTicket_estimatedWaitTime :: Lens' MatchmakingTicket (Maybe Natural)

Average amount of time (in seconds) that players are currently waiting for a match. If there is not enough recent data, this property may be empty.

matchmakingTicket_statusMessage :: Lens' MatchmakingTicket (Maybe Text)

Additional information about the current status.

matchmakingTicket_endTime :: Lens' MatchmakingTicket (Maybe UTCTime)

Time stamp indicating when this matchmaking request stopped being processed due to success, failure, or cancellation. Format is a number expressed in Unix time as milliseconds (for example "1469498468.057").

matchmakingTicket_configurationArn :: Lens' MatchmakingTicket (Maybe Text)

The Amazon Resource Name (ARN) associated with the GameLift matchmaking configuration resource that is used with this ticket.

matchmakingTicket_statusReason :: Lens' MatchmakingTicket (Maybe Text)

Code to explain the current status. For example, a status reason may indicate when a ticket has returned to SEARCHING status after a proposed match fails to receive player acceptances.

matchmakingTicket_players :: Lens' MatchmakingTicket (Maybe [Player])

A set of Player objects, each representing a player to find matches for. Players are identified by a unique player ID and may include latency data for use during matchmaking. If the ticket is in status COMPLETED, the Player objects include the team the players were assigned to in the resulting match.