Amazonka.GameLift.Types.S3Location

Description

Documentation

data S3Location

The location in Amazon S3 where build or script files are stored for access by Amazon GameLift. This location is specified in CreateBuild, CreateScript, and UpdateScript requests.

See: newS3Location smart constructor.

Constructors

S3Location'

Fields

bucket :: Maybe Text

An Amazon S3 bucket identifier. This is the name of the S3 bucket.

GameLift currently does not support uploading from Amazon S3 buckets with names that contain a dot (.).

key :: Maybe Text

The name of the zip file that contains the build files or script files.

objectVersion :: Maybe Text

The version of the file, if object versioning is turned on for the bucket. Amazon GameLift uses this information when retrieving files from an S3 bucket that you own. Use this parameter to specify a specific version of the file. If not set, the latest version of the file is retrieved.

roleArn :: Maybe Text

The Amazon Resource Name (ARN) for an IAM role that allows Amazon GameLift to access the S3 bucket.

newS3Location :: S3Location

Create a value of S3Location with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:bucket:S3Location', s3Location_bucket - An Amazon S3 bucket identifier. This is the name of the S3 bucket.

GameLift currently does not support uploading from Amazon S3 buckets with names that contain a dot (.).

$sel:key:S3Location', s3Location_key - The name of the zip file that contains the build files or script files.

$sel:objectVersion:S3Location', s3Location_objectVersion - The version of the file, if object versioning is turned on for the bucket. Amazon GameLift uses this information when retrieving files from an S3 bucket that you own. Use this parameter to specify a specific version of the file. If not set, the latest version of the file is retrieved.

$sel:roleArn:S3Location', s3Location_roleArn - The Amazon Resource Name (ARN) for an IAM role that allows Amazon GameLift to access the S3 bucket.

s3Location_bucket :: Lens' S3Location (Maybe Text)

An Amazon S3 bucket identifier. This is the name of the S3 bucket.

GameLift currently does not support uploading from Amazon S3 buckets with names that contain a dot (.).

s3Location_key :: Lens' S3Location (Maybe Text)

The name of the zip file that contains the build files or script files.

s3Location_objectVersion :: Lens' S3Location (Maybe Text)

The version of the file, if object versioning is turned on for the bucket. Amazon GameLift uses this information when retrieving files from an S3 bucket that you own. Use this parameter to specify a specific version of the file. If not set, the latest version of the file is retrieved.

s3Location_roleArn :: Lens' S3Location (Maybe Text)

The Amazon Resource Name (ARN) for an IAM role that allows Amazon GameLift to access the S3 bucket.