Amazonka.GameLift.Types.ScalingPolicy

Description

Documentation

data ScalingPolicy

Rule that controls how a fleet is scaled. Scaling policies are uniquely identified by the combination of name and fleet ID.

Related actions

DescribeFleetCapacity | UpdateFleetCapacity | DescribeEC2InstanceLimits | PutScalingPolicy | DescribeScalingPolicies | DeleteScalingPolicy | StopFleetActions | StartFleetActions | All APIs by task

See: newScalingPolicy smart constructor.

Constructors

ScalingPolicy'

Fields

status :: Maybe ScalingStatusType

Current status of the scaling policy. The scaling policy can be in force only when in an ACTIVE status. Scaling policies can be suspended for individual fleets (see StopFleetActions; if suspended for a fleet, the policy status does not change. View a fleet's stopped actions by calling DescribeFleetCapacity.

ACTIVE -- The scaling policy can be used for auto-scaling a fleet.

UPDATE_REQUESTED -- A request to update the scaling policy has been received.

UPDATING -- A change is being made to the scaling policy.

DELETE_REQUESTED -- A request to delete the scaling policy has been received.

DELETING -- The scaling policy is being deleted.

DELETED -- The scaling policy has been deleted.

ERROR -- An error occurred in creating the policy. It should be removed and recreated.

scalingAdjustmentType :: Maybe ScalingAdjustmentType

The type of adjustment to make to a fleet's instance count (see FleetCapacity):

ChangeInCapacity -- add (or subtract) the scaling adjustment value from the current instance count. Positive values scale up while negative values scale down.

ExactCapacity -- set the instance count to the scaling adjustment value.

PercentChangeInCapacity -- increase or reduce the current instance count by the scaling adjustment, read as a percentage. Positive values scale up while negative values scale down.

location :: Maybe Text

evaluationPeriods :: Maybe Natural

Length of time (in minutes) the metric must be at or beyond the threshold before a scaling event is triggered.

policyType :: Maybe PolicyType

The type of scaling policy to create. For a target-based policy, set the parameter MetricName to 'PercentAvailableGameSessions' and specify a TargetConfiguration. For a rule-based policy set the following parameters: MetricName, ComparisonOperator, Threshold, EvaluationPeriods, ScalingAdjustmentType, and ScalingAdjustment.

metricName :: Maybe MetricName

Name of the Amazon GameLift-defined metric that is used to trigger a scaling adjustment. For detailed descriptions of fleet metrics, see Monitor Amazon GameLift with Amazon CloudWatch.

ActivatingGameSessions -- Game sessions in the process of being created.

ActiveGameSessions -- Game sessions that are currently running.

ActiveInstances -- Fleet instances that are currently running at least one game session.

AvailableGameSessions -- Additional game sessions that fleet could host simultaneously, given current capacity.

AvailablePlayerSessions -- Empty player slots in currently active game sessions. This includes game sessions that are not currently accepting players. Reserved player slots are not included.

CurrentPlayerSessions -- Player slots in active game sessions that are being used by a player or are reserved for a player.

IdleInstances -- Active instances that are currently hosting zero game sessions.

PercentAvailableGameSessions -- Unused percentage of the total number of game sessions that a fleet could host simultaneously, given current capacity. Use this metric for a target-based scaling policy.

PercentIdleInstances -- Percentage of the total number of active instances that are hosting zero game sessions.

QueueDepth -- Pending game session placement requests, in any queue, where the current fleet is the top-priority destination.

WaitTime -- Current wait time for pending game session placement requests, in any queue, where the current fleet is the top-priority destination.

fleetArn :: Maybe Text

The Amazon Resource Name (ARN) that is assigned to a GameLift fleet resource and uniquely identifies it. ARNs are unique across all Regions. Format is arn:aws:gamelift:<region>::fleet/fleet-a1234567-b8c9-0d1e-2fa3-b45c6d7e8912.

comparisonOperator :: Maybe ComparisonOperatorType

Comparison operator to use when measuring a metric against the threshold value.

name :: Maybe Text

A descriptive label that is associated with a fleet's scaling policy. Policy names do not need to be unique.

threshold :: Maybe Double

Metric value used to trigger a scaling event.

scalingAdjustment :: Maybe Int

Amount of adjustment to make, based on the scaling adjustment type.

updateStatus :: Maybe LocationUpdateStatus

The current status of the fleet's scaling policies in a requested fleet location. The status PENDING_UPDATE indicates that an update was requested for the fleet but has not yet been completed for the location.

fleetId :: Maybe Text

A unique identifier for the fleet that is associated with this scaling policy.

targetConfiguration :: Maybe TargetConfiguration

An object that contains settings for a target-based scaling policy.

newScalingPolicy :: ScalingPolicy

Create a value of ScalingPolicy with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:status:ScalingPolicy', scalingPolicy_status - Current status of the scaling policy. The scaling policy can be in force only when in an ACTIVE status. Scaling policies can be suspended for individual fleets (see StopFleetActions; if suspended for a fleet, the policy status does not change. View a fleet's stopped actions by calling DescribeFleetCapacity.

ACTIVE -- The scaling policy can be used for auto-scaling a fleet.

UPDATE_REQUESTED -- A request to update the scaling policy has been received.

UPDATING -- A change is being made to the scaling policy.

DELETE_REQUESTED -- A request to delete the scaling policy has been received.

DELETING -- The scaling policy is being deleted.

DELETED -- The scaling policy has been deleted.

ERROR -- An error occurred in creating the policy. It should be removed and recreated.

$sel:scalingAdjustmentType:ScalingPolicy', scalingPolicy_scalingAdjustmentType - The type of adjustment to make to a fleet's instance count (see FleetCapacity):

ChangeInCapacity -- add (or subtract) the scaling adjustment value from the current instance count. Positive values scale up while negative values scale down.

ExactCapacity -- set the instance count to the scaling adjustment value.

PercentChangeInCapacity -- increase or reduce the current instance count by the scaling adjustment, read as a percentage. Positive values scale up while negative values scale down.

$sel:location:ScalingPolicy', scalingPolicy_location -

$sel:evaluationPeriods:ScalingPolicy', scalingPolicy_evaluationPeriods - Length of time (in minutes) the metric must be at or beyond the threshold before a scaling event is triggered.

$sel:policyType:ScalingPolicy', scalingPolicy_policyType - The type of scaling policy to create. For a target-based policy, set the parameter MetricName to 'PercentAvailableGameSessions' and specify a TargetConfiguration. For a rule-based policy set the following parameters: MetricName, ComparisonOperator, Threshold, EvaluationPeriods, ScalingAdjustmentType, and ScalingAdjustment.

$sel:metricName:ScalingPolicy', scalingPolicy_metricName - Name of the Amazon GameLift-defined metric that is used to trigger a scaling adjustment. For detailed descriptions of fleet metrics, see Monitor Amazon GameLift with Amazon CloudWatch.

ActivatingGameSessions -- Game sessions in the process of being created.

ActiveGameSessions -- Game sessions that are currently running.

ActiveInstances -- Fleet instances that are currently running at least one game session.

AvailableGameSessions -- Additional game sessions that fleet could host simultaneously, given current capacity.

AvailablePlayerSessions -- Empty player slots in currently active game sessions. This includes game sessions that are not currently accepting players. Reserved player slots are not included.

CurrentPlayerSessions -- Player slots in active game sessions that are being used by a player or are reserved for a player.

IdleInstances -- Active instances that are currently hosting zero game sessions.

PercentAvailableGameSessions -- Unused percentage of the total number of game sessions that a fleet could host simultaneously, given current capacity. Use this metric for a target-based scaling policy.

PercentIdleInstances -- Percentage of the total number of active instances that are hosting zero game sessions.

QueueDepth -- Pending game session placement requests, in any queue, where the current fleet is the top-priority destination.

WaitTime -- Current wait time for pending game session placement requests, in any queue, where the current fleet is the top-priority destination.

$sel:fleetArn:ScalingPolicy', scalingPolicy_fleetArn - The Amazon Resource Name (ARN) that is assigned to a GameLift fleet resource and uniquely identifies it. ARNs are unique across all Regions. Format is arn:aws:gamelift:<region>::fleet/fleet-a1234567-b8c9-0d1e-2fa3-b45c6d7e8912.

$sel:comparisonOperator:ScalingPolicy', scalingPolicy_comparisonOperator - Comparison operator to use when measuring a metric against the threshold value.

$sel:name:ScalingPolicy', scalingPolicy_name - A descriptive label that is associated with a fleet's scaling policy. Policy names do not need to be unique.

$sel:threshold:ScalingPolicy', scalingPolicy_threshold - Metric value used to trigger a scaling event.

$sel:scalingAdjustment:ScalingPolicy', scalingPolicy_scalingAdjustment - Amount of adjustment to make, based on the scaling adjustment type.

$sel:updateStatus:ScalingPolicy', scalingPolicy_updateStatus - The current status of the fleet's scaling policies in a requested fleet location. The status PENDING_UPDATE indicates that an update was requested for the fleet but has not yet been completed for the location.

$sel:fleetId:ScalingPolicy', scalingPolicy_fleetId - A unique identifier for the fleet that is associated with this scaling policy.

$sel:targetConfiguration:ScalingPolicy', scalingPolicy_targetConfiguration - An object that contains settings for a target-based scaling policy.

scalingPolicy_status :: Lens' ScalingPolicy (Maybe ScalingStatusType)

Current status of the scaling policy. The scaling policy can be in force only when in an ACTIVE status. Scaling policies can be suspended for individual fleets (see StopFleetActions; if suspended for a fleet, the policy status does not change. View a fleet's stopped actions by calling DescribeFleetCapacity.

ACTIVE -- The scaling policy can be used for auto-scaling a fleet.

UPDATE_REQUESTED -- A request to update the scaling policy has been received.

UPDATING -- A change is being made to the scaling policy.

DELETE_REQUESTED -- A request to delete the scaling policy has been received.

DELETING -- The scaling policy is being deleted.

DELETED -- The scaling policy has been deleted.

ERROR -- An error occurred in creating the policy. It should be removed and recreated.

scalingPolicy_scalingAdjustmentType :: Lens' ScalingPolicy (Maybe ScalingAdjustmentType)

The type of adjustment to make to a fleet's instance count (see FleetCapacity):

ChangeInCapacity -- add (or subtract) the scaling adjustment value from the current instance count. Positive values scale up while negative values scale down.

ExactCapacity -- set the instance count to the scaling adjustment value.

PercentChangeInCapacity -- increase or reduce the current instance count by the scaling adjustment, read as a percentage. Positive values scale up while negative values scale down.

scalingPolicy_location :: Lens' ScalingPolicy (Maybe Text)

scalingPolicy_evaluationPeriods :: Lens' ScalingPolicy (Maybe Natural)

Length of time (in minutes) the metric must be at or beyond the threshold before a scaling event is triggered.

scalingPolicy_policyType :: Lens' ScalingPolicy (Maybe PolicyType)

The type of scaling policy to create. For a target-based policy, set the parameter MetricName to 'PercentAvailableGameSessions' and specify a TargetConfiguration. For a rule-based policy set the following parameters: MetricName, ComparisonOperator, Threshold, EvaluationPeriods, ScalingAdjustmentType, and ScalingAdjustment.

scalingPolicy_metricName :: Lens' ScalingPolicy (Maybe MetricName)

Name of the Amazon GameLift-defined metric that is used to trigger a scaling adjustment. For detailed descriptions of fleet metrics, see Monitor Amazon GameLift with Amazon CloudWatch.

ActivatingGameSessions -- Game sessions in the process of being created.

ActiveGameSessions -- Game sessions that are currently running.

ActiveInstances -- Fleet instances that are currently running at least one game session.

AvailableGameSessions -- Additional game sessions that fleet could host simultaneously, given current capacity.

AvailablePlayerSessions -- Empty player slots in currently active game sessions. This includes game sessions that are not currently accepting players. Reserved player slots are not included.

CurrentPlayerSessions -- Player slots in active game sessions that are being used by a player or are reserved for a player.

IdleInstances -- Active instances that are currently hosting zero game sessions.

PercentAvailableGameSessions -- Unused percentage of the total number of game sessions that a fleet could host simultaneously, given current capacity. Use this metric for a target-based scaling policy.

PercentIdleInstances -- Percentage of the total number of active instances that are hosting zero game sessions.

QueueDepth -- Pending game session placement requests, in any queue, where the current fleet is the top-priority destination.

WaitTime -- Current wait time for pending game session placement requests, in any queue, where the current fleet is the top-priority destination.

scalingPolicy_fleetArn :: Lens' ScalingPolicy (Maybe Text)

The Amazon Resource Name (ARN) that is assigned to a GameLift fleet resource and uniquely identifies it. ARNs are unique across all Regions. Format is arn:aws:gamelift:<region>::fleet/fleet-a1234567-b8c9-0d1e-2fa3-b45c6d7e8912.

scalingPolicy_comparisonOperator :: Lens' ScalingPolicy (Maybe ComparisonOperatorType)

Comparison operator to use when measuring a metric against the threshold value.

scalingPolicy_name :: Lens' ScalingPolicy (Maybe Text)

A descriptive label that is associated with a fleet's scaling policy. Policy names do not need to be unique.

scalingPolicy_threshold :: Lens' ScalingPolicy (Maybe Double)

Metric value used to trigger a scaling event.

scalingPolicy_scalingAdjustment :: Lens' ScalingPolicy (Maybe Int)

Amount of adjustment to make, based on the scaling adjustment type.

scalingPolicy_updateStatus :: Lens' ScalingPolicy (Maybe LocationUpdateStatus)

The current status of the fleet's scaling policies in a requested fleet location. The status PENDING_UPDATE indicates that an update was requested for the fleet but has not yet been completed for the location.

scalingPolicy_fleetId :: Lens' ScalingPolicy (Maybe Text)

A unique identifier for the fleet that is associated with this scaling policy.

scalingPolicy_targetConfiguration :: Lens' ScalingPolicy (Maybe TargetConfiguration)

An object that contains settings for a target-based scaling policy.