Amazonka.GameLift.Types.VpcPeeringAuthorization

Description

Documentation

data VpcPeeringAuthorization

Represents an authorization for a VPC peering connection between the VPC for an Amazon GameLift fleet and another VPC on an account you have access to. This authorization must exist and be valid for the peering connection to be established. Authorizations are valid for 24 hours after they are issued.

Related actions

CreateVpcPeeringAuthorization | DescribeVpcPeeringAuthorizations | DeleteVpcPeeringAuthorization | CreateVpcPeeringConnection | DescribeVpcPeeringConnections | DeleteVpcPeeringConnection | All APIs by task

See: newVpcPeeringAuthorization smart constructor.

Constructors

VpcPeeringAuthorization'

Fields

creationTime :: Maybe POSIX

Time stamp indicating when this authorization was issued. Format is a number expressed in Unix time as milliseconds (for example "1469498468.057").

peerVpcId :: Maybe Text

A unique identifier for a VPC with resources to be accessed by your GameLift fleet. The VPC must be in the same Region as your fleet. To look up a VPC ID, use the VPC Dashboard in the AWS Management Console. Learn more about VPC peering in VPC Peering with GameLift Fleets.

peerVpcAwsAccountId :: Maybe Text

gameLiftAwsAccountId :: Maybe Text

A unique identifier for the AWS account that you use to manage your GameLift fleet. You can find your Account ID in the AWS Management Console under account settings.

expirationTime :: Maybe POSIX

Time stamp indicating when this authorization expires (24 hours after issuance). Format is a number expressed in Unix time as milliseconds (for example "1469498468.057").

newVpcPeeringAuthorization :: VpcPeeringAuthorization

Create a value of VpcPeeringAuthorization with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:creationTime:VpcPeeringAuthorization', vpcPeeringAuthorization_creationTime - Time stamp indicating when this authorization was issued. Format is a number expressed in Unix time as milliseconds (for example "1469498468.057").

$sel:peerVpcId:VpcPeeringAuthorization', vpcPeeringAuthorization_peerVpcId - A unique identifier for a VPC with resources to be accessed by your GameLift fleet. The VPC must be in the same Region as your fleet. To look up a VPC ID, use the VPC Dashboard in the AWS Management Console. Learn more about VPC peering in VPC Peering with GameLift Fleets.

$sel:peerVpcAwsAccountId:VpcPeeringAuthorization', vpcPeeringAuthorization_peerVpcAwsAccountId -

$sel:gameLiftAwsAccountId:VpcPeeringAuthorization', vpcPeeringAuthorization_gameLiftAwsAccountId - A unique identifier for the AWS account that you use to manage your GameLift fleet. You can find your Account ID in the AWS Management Console under account settings.

$sel:expirationTime:VpcPeeringAuthorization', vpcPeeringAuthorization_expirationTime - Time stamp indicating when this authorization expires (24 hours after issuance). Format is a number expressed in Unix time as milliseconds (for example "1469498468.057").

vpcPeeringAuthorization_creationTime :: Lens' VpcPeeringAuthorization (Maybe UTCTime)

Time stamp indicating when this authorization was issued. Format is a number expressed in Unix time as milliseconds (for example "1469498468.057").

vpcPeeringAuthorization_peerVpcId :: Lens' VpcPeeringAuthorization (Maybe Text)

A unique identifier for a VPC with resources to be accessed by your GameLift fleet. The VPC must be in the same Region as your fleet. To look up a VPC ID, use the VPC Dashboard in the AWS Management Console. Learn more about VPC peering in VPC Peering with GameLift Fleets.

vpcPeeringAuthorization_peerVpcAwsAccountId :: Lens' VpcPeeringAuthorization (Maybe Text)

vpcPeeringAuthorization_gameLiftAwsAccountId :: Lens' VpcPeeringAuthorization (Maybe Text)

A unique identifier for the AWS account that you use to manage your GameLift fleet. You can find your Account ID in the AWS Management Console under account settings.

vpcPeeringAuthorization_expirationTime :: Lens' VpcPeeringAuthorization (Maybe UTCTime)

Time stamp indicating when this authorization expires (24 hours after issuance). Format is a number expressed in Unix time as milliseconds (for example "1469498468.057").