Amazonka.GameLift.UpdateGameSession

Description

Updates the mutable properties of a game session.

To update a game session, specify the game session ID and the values you want to change.

If successful, the updated GameSession object is returned.

Related actions

CreateGameSession | DescribeGameSessions | DescribeGameSessionDetails | SearchGameSessions | UpdateGameSession | GetGameSessionLogUrl | StartGameSessionPlacement | DescribeGameSessionPlacement | StopGameSessionPlacement | All APIs by task

Creating a Request

data UpdateGameSession

Represents the input for a request operation.

See: newUpdateGameSession smart constructor.

Constructors

UpdateGameSession'

Fields

maximumPlayerSessionCount :: Maybe Natural

The maximum number of players that can be connected simultaneously to the game session.

playerSessionCreationPolicy :: Maybe PlayerSessionCreationPolicy

A policy that determines whether the game session is accepting new players.

name :: Maybe Text

A descriptive label that is associated with a game session. Session names do not need to be unique.

protectionPolicy :: Maybe ProtectionPolicy

Game session protection policy to apply to this game session only.

NoProtection -- The game session can be terminated during a scale-down event.

FullProtection -- If the game session is in an ACTIVE status, it cannot be terminated during a scale-down event.

gameSessionId :: Text

A unique identifier for the game session to update.

newUpdateGameSession

Arguments

:: Text
$sel:gameSessionId:UpdateGameSession'

-> UpdateGameSession

Create a value of UpdateGameSession with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:maximumPlayerSessionCount:UpdateGameSession', updateGameSession_maximumPlayerSessionCount - The maximum number of players that can be connected simultaneously to the game session.

$sel:playerSessionCreationPolicy:UpdateGameSession', updateGameSession_playerSessionCreationPolicy - A policy that determines whether the game session is accepting new players.

$sel:name:UpdateGameSession', updateGameSession_name - A descriptive label that is associated with a game session. Session names do not need to be unique.

$sel:protectionPolicy:UpdateGameSession', updateGameSession_protectionPolicy - Game session protection policy to apply to this game session only.

NoProtection -- The game session can be terminated during a scale-down event.

FullProtection -- If the game session is in an ACTIVE status, it cannot be terminated during a scale-down event.

$sel:gameSessionId:UpdateGameSession', updateGameSession_gameSessionId - A unique identifier for the game session to update.

Request Lenses

updateGameSession_maximumPlayerSessionCount :: Lens' UpdateGameSession (Maybe Natural)

The maximum number of players that can be connected simultaneously to the game session.

updateGameSession_playerSessionCreationPolicy :: Lens' UpdateGameSession (Maybe PlayerSessionCreationPolicy)

A policy that determines whether the game session is accepting new players.

updateGameSession_name :: Lens' UpdateGameSession (Maybe Text)

A descriptive label that is associated with a game session. Session names do not need to be unique.

updateGameSession_protectionPolicy :: Lens' UpdateGameSession (Maybe ProtectionPolicy)

Game session protection policy to apply to this game session only.

NoProtection -- The game session can be terminated during a scale-down event.

FullProtection -- If the game session is in an ACTIVE status, it cannot be terminated during a scale-down event.

updateGameSession_gameSessionId :: Lens' UpdateGameSession Text

A unique identifier for the game session to update.

Destructuring the Response

data UpdateGameSessionResponse

Represents the returned data in response to a request operation.

See: newUpdateGameSessionResponse smart constructor.

Constructors

UpdateGameSessionResponse'

Fields

gameSession :: Maybe GameSession

The updated game session properties.

httpStatus :: Int

The response's http status code.

newUpdateGameSessionResponse

Arguments

:: Int
$sel:httpStatus:UpdateGameSessionResponse'

-> UpdateGameSessionResponse

Create a value of UpdateGameSessionResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:gameSession:UpdateGameSessionResponse', updateGameSessionResponse_gameSession - The updated game session properties.

$sel:httpStatus:UpdateGameSessionResponse', updateGameSessionResponse_httpStatus - The response's http status code.

Response Lenses

updateGameSessionResponse_gameSession :: Lens' UpdateGameSessionResponse (Maybe GameSession)

The updated game session properties.

updateGameSessionResponse_httpStatus :: Lens' UpdateGameSessionResponse Int

The response's http status code.