Amazonka.GameLift.UpdateGameSessionQueue

Description

Updates the configuration of a game session queue, which determines how the queue processes new game session requests. To update settings, specify the queue name to be updated and provide the new settings. When updating destinations, provide a complete list of destinations.

Learn more

Using Multi-Region Queues

Related actions

CreateGameSessionQueue | DescribeGameSessionQueues | UpdateGameSessionQueue | DeleteGameSessionQueue | All APIs by task

Creating a Request

data UpdateGameSessionQueue

Represents the input for a request operation.

See: newUpdateGameSessionQueue smart constructor.

Constructors

UpdateGameSessionQueue'

Fields

playerLatencyPolicies :: Maybe [PlayerLatencyPolicy]

A set of policies that act as a sliding cap on player latency. FleetIQ works to deliver low latency for most players in a game session. These policies ensure that no individual player can be placed into a game with unreasonably high latency. Use multiple policies to gradually relax latency requirements a step at a time. Multiple policies are applied based on their maximum allowed latency, starting with the lowest value. When updating policies, provide a complete collection of policies.

filterConfiguration :: Maybe FilterConfiguration

A list of locations where a queue is allowed to place new game sessions. Locations are specified in the form of AWS Region codes, such as us-west-2. If this parameter is not set, game sessions can be placed in any queue location. To remove an existing filter configuration, pass in an empty set.

notificationTarget :: Maybe Text

An SNS topic ARN that is set up to receive game session placement notifications. See Setting up notifications for game session placement.

timeoutInSeconds :: Maybe Natural

The maximum time, in seconds, that a new game session placement request remains in the queue. When a request exceeds this time, the game session placement changes to a TIMED_OUT status.

destinations :: Maybe [GameSessionQueueDestination]

A list of fleets and/or fleet aliases that can be used to fulfill game session placement requests in the queue. Destinations are identified by either a fleet ARN or a fleet alias ARN, and are listed in order of placement preference. When updating this list, provide a complete list of destinations.

customEventData :: Maybe Text

Information to be added to all events that are related to this game session queue.

priorityConfiguration :: Maybe PriorityConfiguration

Custom settings to use when prioritizing destinations and locations for game session placements. This configuration replaces the FleetIQ default prioritization process. Priority types that are not explicitly named will be automatically applied at the end of the prioritization process. To remove an existing priority configuration, pass in an empty set.

name :: Text

A descriptive label that is associated with game session queue. Queue names must be unique within each Region. You can use either the queue ID or ARN value.

newUpdateGameSessionQueue

Arguments

:: Text
$sel:name:UpdateGameSessionQueue'

-> UpdateGameSessionQueue

Create a value of UpdateGameSessionQueue with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:playerLatencyPolicies:UpdateGameSessionQueue', updateGameSessionQueue_playerLatencyPolicies - A set of policies that act as a sliding cap on player latency. FleetIQ works to deliver low latency for most players in a game session. These policies ensure that no individual player can be placed into a game with unreasonably high latency. Use multiple policies to gradually relax latency requirements a step at a time. Multiple policies are applied based on their maximum allowed latency, starting with the lowest value. When updating policies, provide a complete collection of policies.

$sel:filterConfiguration:UpdateGameSessionQueue', updateGameSessionQueue_filterConfiguration - A list of locations where a queue is allowed to place new game sessions. Locations are specified in the form of AWS Region codes, such as us-west-2. If this parameter is not set, game sessions can be placed in any queue location. To remove an existing filter configuration, pass in an empty set.

$sel:notificationTarget:UpdateGameSessionQueue', updateGameSessionQueue_notificationTarget - An SNS topic ARN that is set up to receive game session placement notifications. See Setting up notifications for game session placement.

$sel:timeoutInSeconds:UpdateGameSessionQueue', updateGameSessionQueue_timeoutInSeconds - The maximum time, in seconds, that a new game session placement request remains in the queue. When a request exceeds this time, the game session placement changes to a TIMED_OUT status.

$sel:destinations:UpdateGameSessionQueue', updateGameSessionQueue_destinations - A list of fleets and/or fleet aliases that can be used to fulfill game session placement requests in the queue. Destinations are identified by either a fleet ARN or a fleet alias ARN, and are listed in order of placement preference. When updating this list, provide a complete list of destinations.

$sel:customEventData:UpdateGameSessionQueue', updateGameSessionQueue_customEventData - Information to be added to all events that are related to this game session queue.

$sel:priorityConfiguration:UpdateGameSessionQueue', updateGameSessionQueue_priorityConfiguration - Custom settings to use when prioritizing destinations and locations for game session placements. This configuration replaces the FleetIQ default prioritization process. Priority types that are not explicitly named will be automatically applied at the end of the prioritization process. To remove an existing priority configuration, pass in an empty set.

$sel:name:UpdateGameSessionQueue', updateGameSessionQueue_name - A descriptive label that is associated with game session queue. Queue names must be unique within each Region. You can use either the queue ID or ARN value.

Request Lenses

updateGameSessionQueue_playerLatencyPolicies :: Lens' UpdateGameSessionQueue (Maybe [PlayerLatencyPolicy])

A set of policies that act as a sliding cap on player latency. FleetIQ works to deliver low latency for most players in a game session. These policies ensure that no individual player can be placed into a game with unreasonably high latency. Use multiple policies to gradually relax latency requirements a step at a time. Multiple policies are applied based on their maximum allowed latency, starting with the lowest value. When updating policies, provide a complete collection of policies.

updateGameSessionQueue_filterConfiguration :: Lens' UpdateGameSessionQueue (Maybe FilterConfiguration)

A list of locations where a queue is allowed to place new game sessions. Locations are specified in the form of AWS Region codes, such as us-west-2. If this parameter is not set, game sessions can be placed in any queue location. To remove an existing filter configuration, pass in an empty set.

updateGameSessionQueue_notificationTarget :: Lens' UpdateGameSessionQueue (Maybe Text)

An SNS topic ARN that is set up to receive game session placement notifications. See Setting up notifications for game session placement.

updateGameSessionQueue_timeoutInSeconds :: Lens' UpdateGameSessionQueue (Maybe Natural)

The maximum time, in seconds, that a new game session placement request remains in the queue. When a request exceeds this time, the game session placement changes to a TIMED_OUT status.

updateGameSessionQueue_destinations :: Lens' UpdateGameSessionQueue (Maybe [GameSessionQueueDestination])

A list of fleets and/or fleet aliases that can be used to fulfill game session placement requests in the queue. Destinations are identified by either a fleet ARN or a fleet alias ARN, and are listed in order of placement preference. When updating this list, provide a complete list of destinations.

updateGameSessionQueue_customEventData :: Lens' UpdateGameSessionQueue (Maybe Text)

Information to be added to all events that are related to this game session queue.

updateGameSessionQueue_priorityConfiguration :: Lens' UpdateGameSessionQueue (Maybe PriorityConfiguration)

Custom settings to use when prioritizing destinations and locations for game session placements. This configuration replaces the FleetIQ default prioritization process. Priority types that are not explicitly named will be automatically applied at the end of the prioritization process. To remove an existing priority configuration, pass in an empty set.

updateGameSessionQueue_name :: Lens' UpdateGameSessionQueue Text

A descriptive label that is associated with game session queue. Queue names must be unique within each Region. You can use either the queue ID or ARN value.

Destructuring the Response

data UpdateGameSessionQueueResponse

Represents the returned data in response to a request operation.

See: newUpdateGameSessionQueueResponse smart constructor.

Constructors

UpdateGameSessionQueueResponse'

Fields

gameSessionQueue :: Maybe GameSessionQueue

An object that describes the newly updated game session queue.

httpStatus :: Int

The response's http status code.

newUpdateGameSessionQueueResponse

Arguments

:: Int
$sel:httpStatus:UpdateGameSessionQueueResponse'

-> UpdateGameSessionQueueResponse

Create a value of UpdateGameSessionQueueResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:gameSessionQueue:UpdateGameSessionQueueResponse', updateGameSessionQueueResponse_gameSessionQueue - An object that describes the newly updated game session queue.

$sel:httpStatus:UpdateGameSessionQueueResponse', updateGameSessionQueueResponse_httpStatus - The response's http status code.

Response Lenses

updateGameSessionQueueResponse_gameSessionQueue :: Lens' UpdateGameSessionQueueResponse (Maybe GameSessionQueue)

An object that describes the newly updated game session queue.

updateGameSessionQueueResponse_httpStatus :: Lens' UpdateGameSessionQueueResponse Int

The response's http status code.