Amazonka.GameLift.UpdateMatchmakingConfiguration

Description

Updates settings for a FlexMatch matchmaking configuration. These changes affect all matches and game sessions that are created after the update. To update settings, specify the configuration name to be updated and provide the new settings.

Learn more

Design a FlexMatch matchmaker

Related actions

CreateMatchmakingConfiguration | DescribeMatchmakingConfigurations | UpdateMatchmakingConfiguration | DeleteMatchmakingConfiguration | CreateMatchmakingRuleSet | DescribeMatchmakingRuleSets | ValidateMatchmakingRuleSet | DeleteMatchmakingRuleSet | All APIs by task

Creating a Request

data UpdateMatchmakingConfiguration

Represents the input for a request operation.

See: newUpdateMatchmakingConfiguration smart constructor.

Constructors

UpdateMatchmakingConfiguration'

Fields

backfillMode :: Maybe BackfillMode

The method that is used to backfill game sessions created with this matchmaking configuration. Specify MANUAL when your game manages backfill requests manually or does not use the match backfill feature. Specify AUTOMATIC to have GameLift create a StartMatchBackfill request whenever a game session has one or more open slots. Learn more about manual and automatic backfill in Backfill Existing Games with FlexMatch. Automatic backfill is not available when FlexMatchMode is set to STANDALONE.

gameProperties :: Maybe [GameProperty]

A set of custom properties for a game session, formatted as key:value pairs. These properties are passed to a game server process in the GameSession object with a request to start a new game session (see Start a Game Session). This information is added to the new GameSession object that is created for a successful match. This parameter is not used if FlexMatchMode is set to STANDALONE.

ruleSetName :: Maybe Text

A unique identifier for the matchmaking rule set to use with this configuration. You can use either the rule set name or ARN value. A matchmaking configuration can only use rule sets that are defined in the same Region.

acceptanceTimeoutSeconds :: Maybe Natural

The length of time (in seconds) to wait for players to accept a proposed match, if acceptance is required.

requestTimeoutSeconds :: Maybe Natural

The maximum duration, in seconds, that a matchmaking ticket can remain in process before timing out. Requests that fail due to timing out can be resubmitted as needed.

notificationTarget :: Maybe Text

An SNS topic ARN that is set up to receive matchmaking notifications. See Setting up notifications for matchmaking for more information.

flexMatchMode :: Maybe FlexMatchMode

Indicates whether this matchmaking configuration is being used with GameLift hosting or as a standalone matchmaking solution.

STANDALONE - FlexMatch forms matches and returns match information, including players and team assignments, in a MatchmakingSucceeded event.

WITH_QUEUE - FlexMatch forms matches and uses the specified GameLift queue to start a game session for the match.

gameSessionQueueArns :: Maybe [Text]

The Amazon Resource Name (ARN) that is assigned to a GameLift game session queue resource and uniquely identifies it. ARNs are unique across all Regions. Format is arn:aws:gamelift:<region>::gamesessionqueue/<queue name>. Queues can be located in any Region. Queues are used to start new GameLift-hosted game sessions for matches that are created with this matchmaking configuration. If FlexMatchMode is set to STANDALONE, do not set this parameter.

customEventData :: Maybe Text

Information to add to all events related to the matchmaking configuration.

acceptanceRequired :: Maybe Bool

A flag that indicates whether a match that was created with this configuration must be accepted by the matched players. To require acceptance, set to TRUE. With this option enabled, matchmaking tickets use the status REQUIRES_ACCEPTANCE to indicate when a completed potential match is waiting for player acceptance.

gameSessionData :: Maybe Text

A set of custom game session properties, formatted as a single string value. This data is passed to a game server process in the GameSession object with a request to start a new game session (see Start a Game Session). This information is added to the new GameSession object that is created for a successful match. This parameter is not used if FlexMatchMode is set to STANDALONE.

description :: Maybe Text

A descriptive label that is associated with matchmaking configuration.

additionalPlayerCount :: Maybe Natural

The number of player slots in a match to keep open for future players. For example, if the configuration's rule set specifies a match for a single 12-person team, and the additional player count is set to 2, only 10 players are selected for the match. This parameter is not used if FlexMatchMode is set to STANDALONE.

name :: Text

A unique identifier for the matchmaking configuration to update. You can use either the configuration name or ARN value.

newUpdateMatchmakingConfiguration

Arguments

:: Text
$sel:name:UpdateMatchmakingConfiguration'

-> UpdateMatchmakingConfiguration

Create a value of UpdateMatchmakingConfiguration with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:backfillMode:UpdateMatchmakingConfiguration', updateMatchmakingConfiguration_backfillMode - The method that is used to backfill game sessions created with this matchmaking configuration. Specify MANUAL when your game manages backfill requests manually or does not use the match backfill feature. Specify AUTOMATIC to have GameLift create a StartMatchBackfill request whenever a game session has one or more open slots. Learn more about manual and automatic backfill in Backfill Existing Games with FlexMatch. Automatic backfill is not available when FlexMatchMode is set to STANDALONE.

$sel:gameProperties:UpdateMatchmakingConfiguration', updateMatchmakingConfiguration_gameProperties - A set of custom properties for a game session, formatted as key:value pairs. These properties are passed to a game server process in the GameSession object with a request to start a new game session (see Start a Game Session). This information is added to the new GameSession object that is created for a successful match. This parameter is not used if FlexMatchMode is set to STANDALONE.

$sel:ruleSetName:UpdateMatchmakingConfiguration', updateMatchmakingConfiguration_ruleSetName - A unique identifier for the matchmaking rule set to use with this configuration. You can use either the rule set name or ARN value. A matchmaking configuration can only use rule sets that are defined in the same Region.

$sel:acceptanceTimeoutSeconds:UpdateMatchmakingConfiguration', updateMatchmakingConfiguration_acceptanceTimeoutSeconds - The length of time (in seconds) to wait for players to accept a proposed match, if acceptance is required.

$sel:requestTimeoutSeconds:UpdateMatchmakingConfiguration', updateMatchmakingConfiguration_requestTimeoutSeconds - The maximum duration, in seconds, that a matchmaking ticket can remain in process before timing out. Requests that fail due to timing out can be resubmitted as needed.

$sel:notificationTarget:UpdateMatchmakingConfiguration', updateMatchmakingConfiguration_notificationTarget - An SNS topic ARN that is set up to receive matchmaking notifications. See Setting up notifications for matchmaking for more information.

$sel:flexMatchMode:UpdateMatchmakingConfiguration', updateMatchmakingConfiguration_flexMatchMode - Indicates whether this matchmaking configuration is being used with GameLift hosting or as a standalone matchmaking solution.

STANDALONE - FlexMatch forms matches and returns match information, including players and team assignments, in a MatchmakingSucceeded event.

WITH_QUEUE - FlexMatch forms matches and uses the specified GameLift queue to start a game session for the match.

$sel:gameSessionQueueArns:UpdateMatchmakingConfiguration', updateMatchmakingConfiguration_gameSessionQueueArns - The Amazon Resource Name (ARN) that is assigned to a GameLift game session queue resource and uniquely identifies it. ARNs are unique across all Regions. Format is arn:aws:gamelift:<region>::gamesessionqueue/<queue name>. Queues can be located in any Region. Queues are used to start new GameLift-hosted game sessions for matches that are created with this matchmaking configuration. If FlexMatchMode is set to STANDALONE, do not set this parameter.

$sel:customEventData:UpdateMatchmakingConfiguration', updateMatchmakingConfiguration_customEventData - Information to add to all events related to the matchmaking configuration.

$sel:acceptanceRequired:UpdateMatchmakingConfiguration', updateMatchmakingConfiguration_acceptanceRequired - A flag that indicates whether a match that was created with this configuration must be accepted by the matched players. To require acceptance, set to TRUE. With this option enabled, matchmaking tickets use the status REQUIRES_ACCEPTANCE to indicate when a completed potential match is waiting for player acceptance.

$sel:gameSessionData:UpdateMatchmakingConfiguration', updateMatchmakingConfiguration_gameSessionData - A set of custom game session properties, formatted as a single string value. This data is passed to a game server process in the GameSession object with a request to start a new game session (see Start a Game Session). This information is added to the new GameSession object that is created for a successful match. This parameter is not used if FlexMatchMode is set to STANDALONE.

$sel:description:UpdateMatchmakingConfiguration', updateMatchmakingConfiguration_description - A descriptive label that is associated with matchmaking configuration.

$sel:additionalPlayerCount:UpdateMatchmakingConfiguration', updateMatchmakingConfiguration_additionalPlayerCount - The number of player slots in a match to keep open for future players. For example, if the configuration's rule set specifies a match for a single 12-person team, and the additional player count is set to 2, only 10 players are selected for the match. This parameter is not used if FlexMatchMode is set to STANDALONE.

$sel:name:UpdateMatchmakingConfiguration', updateMatchmakingConfiguration_name - A unique identifier for the matchmaking configuration to update. You can use either the configuration name or ARN value.

Request Lenses

updateMatchmakingConfiguration_backfillMode :: Lens' UpdateMatchmakingConfiguration (Maybe BackfillMode)

The method that is used to backfill game sessions created with this matchmaking configuration. Specify MANUAL when your game manages backfill requests manually or does not use the match backfill feature. Specify AUTOMATIC to have GameLift create a StartMatchBackfill request whenever a game session has one or more open slots. Learn more about manual and automatic backfill in Backfill Existing Games with FlexMatch. Automatic backfill is not available when FlexMatchMode is set to STANDALONE.

updateMatchmakingConfiguration_gameProperties :: Lens' UpdateMatchmakingConfiguration (Maybe [GameProperty])

A set of custom properties for a game session, formatted as key:value pairs. These properties are passed to a game server process in the GameSession object with a request to start a new game session (see Start a Game Session). This information is added to the new GameSession object that is created for a successful match. This parameter is not used if FlexMatchMode is set to STANDALONE.

updateMatchmakingConfiguration_ruleSetName :: Lens' UpdateMatchmakingConfiguration (Maybe Text)

A unique identifier for the matchmaking rule set to use with this configuration. You can use either the rule set name or ARN value. A matchmaking configuration can only use rule sets that are defined in the same Region.

updateMatchmakingConfiguration_acceptanceTimeoutSeconds :: Lens' UpdateMatchmakingConfiguration (Maybe Natural)

The length of time (in seconds) to wait for players to accept a proposed match, if acceptance is required.

updateMatchmakingConfiguration_requestTimeoutSeconds :: Lens' UpdateMatchmakingConfiguration (Maybe Natural)

The maximum duration, in seconds, that a matchmaking ticket can remain in process before timing out. Requests that fail due to timing out can be resubmitted as needed.

updateMatchmakingConfiguration_notificationTarget :: Lens' UpdateMatchmakingConfiguration (Maybe Text)

An SNS topic ARN that is set up to receive matchmaking notifications. See Setting up notifications for matchmaking for more information.

updateMatchmakingConfiguration_flexMatchMode :: Lens' UpdateMatchmakingConfiguration (Maybe FlexMatchMode)

Indicates whether this matchmaking configuration is being used with GameLift hosting or as a standalone matchmaking solution.

STANDALONE - FlexMatch forms matches and returns match information, including players and team assignments, in a MatchmakingSucceeded event.

WITH_QUEUE - FlexMatch forms matches and uses the specified GameLift queue to start a game session for the match.

updateMatchmakingConfiguration_gameSessionQueueArns :: Lens' UpdateMatchmakingConfiguration (Maybe [Text])

The Amazon Resource Name (ARN) that is assigned to a GameLift game session queue resource and uniquely identifies it. ARNs are unique across all Regions. Format is arn:aws:gamelift:<region>::gamesessionqueue/<queue name>. Queues can be located in any Region. Queues are used to start new GameLift-hosted game sessions for matches that are created with this matchmaking configuration. If FlexMatchMode is set to STANDALONE, do not set this parameter.

updateMatchmakingConfiguration_customEventData :: Lens' UpdateMatchmakingConfiguration (Maybe Text)

Information to add to all events related to the matchmaking configuration.

updateMatchmakingConfiguration_acceptanceRequired :: Lens' UpdateMatchmakingConfiguration (Maybe Bool)

A flag that indicates whether a match that was created with this configuration must be accepted by the matched players. To require acceptance, set to TRUE. With this option enabled, matchmaking tickets use the status REQUIRES_ACCEPTANCE to indicate when a completed potential match is waiting for player acceptance.

updateMatchmakingConfiguration_gameSessionData :: Lens' UpdateMatchmakingConfiguration (Maybe Text)

A set of custom game session properties, formatted as a single string value. This data is passed to a game server process in the GameSession object with a request to start a new game session (see Start a Game Session). This information is added to the new GameSession object that is created for a successful match. This parameter is not used if FlexMatchMode is set to STANDALONE.

updateMatchmakingConfiguration_description :: Lens' UpdateMatchmakingConfiguration (Maybe Text)

A descriptive label that is associated with matchmaking configuration.

updateMatchmakingConfiguration_additionalPlayerCount :: Lens' UpdateMatchmakingConfiguration (Maybe Natural)

The number of player slots in a match to keep open for future players. For example, if the configuration's rule set specifies a match for a single 12-person team, and the additional player count is set to 2, only 10 players are selected for the match. This parameter is not used if FlexMatchMode is set to STANDALONE.

updateMatchmakingConfiguration_name :: Lens' UpdateMatchmakingConfiguration Text

A unique identifier for the matchmaking configuration to update. You can use either the configuration name or ARN value.

Destructuring the Response

data UpdateMatchmakingConfigurationResponse

Represents the returned data in response to a request operation.

See: newUpdateMatchmakingConfigurationResponse smart constructor.

Constructors

UpdateMatchmakingConfigurationResponse'

Fields

configuration :: Maybe MatchmakingConfiguration

The updated matchmaking configuration.

httpStatus :: Int

The response's http status code.

newUpdateMatchmakingConfigurationResponse

Arguments

:: Int
$sel:httpStatus:UpdateMatchmakingConfigurationResponse'

-> UpdateMatchmakingConfigurationResponse

Create a value of UpdateMatchmakingConfigurationResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:configuration:UpdateMatchmakingConfigurationResponse', updateMatchmakingConfigurationResponse_configuration - The updated matchmaking configuration.

$sel:httpStatus:UpdateMatchmakingConfigurationResponse', updateMatchmakingConfigurationResponse_httpStatus - The response's http status code.

Response Lenses

updateMatchmakingConfigurationResponse_configuration :: Lens' UpdateMatchmakingConfigurationResponse (Maybe MatchmakingConfiguration)

The updated matchmaking configuration.

updateMatchmakingConfigurationResponse_httpStatus :: Lens' UpdateMatchmakingConfigurationResponse Int

The response's http status code.