Amazonka.GameLift.UpdateRuntimeConfiguration

Description

Updates the current runtime configuration for the specified fleet, which tells GameLift how to launch server processes on all instances in the fleet. You can update a fleet's runtime configuration at any time after the fleet is created; it does not need to be in ACTIVE status.

To update runtime configuration, specify the fleet ID and provide a RuntimeConfiguration with an updated set of server process configurations.

If successful, the fleet's runtime configuration settings are updated. Each instance in the fleet regularly checks for and retrieves updated runtime configurations. Instances immediately begin complying with the new configuration by launching new server processes or not replacing existing processes when they shut down. Updating a fleet's runtime configuration never affects existing server processes.

Learn more

Setting up GameLift fleets

Related actions

CreateFleetLocations | UpdateFleetAttributes | UpdateFleetCapacity | UpdateFleetPortSettings | UpdateRuntimeConfiguration | StopFleetActions | StartFleetActions | PutScalingPolicy | DeleteFleet | DeleteFleetLocations | DeleteScalingPolicy | All APIs by task

Creating a Request

data UpdateRuntimeConfiguration

Represents the input for a request operation.

See: newUpdateRuntimeConfiguration smart constructor.

Constructors

UpdateRuntimeConfiguration'

Fields

fleetId :: Text

A unique identifier for the fleet to update runtime configuration for. You can use either the fleet ID or ARN value.

runtimeConfiguration :: RuntimeConfiguration

Instructions for launching server processes on each instance in the fleet. Server processes run either a custom game build executable or a Realtime Servers script. The runtime configuration lists the types of server processes to run on an instance, how to launch them, and the number of processes to run concurrently.

newUpdateRuntimeConfiguration

Arguments

:: Text
$sel:fleetId:UpdateRuntimeConfiguration'

-> RuntimeConfiguration
$sel:runtimeConfiguration:UpdateRuntimeConfiguration'

-> UpdateRuntimeConfiguration

Create a value of UpdateRuntimeConfiguration with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:fleetId:UpdateRuntimeConfiguration', updateRuntimeConfiguration_fleetId - A unique identifier for the fleet to update runtime configuration for. You can use either the fleet ID or ARN value.

$sel:runtimeConfiguration:UpdateRuntimeConfiguration', updateRuntimeConfiguration_runtimeConfiguration - Instructions for launching server processes on each instance in the fleet. Server processes run either a custom game build executable or a Realtime Servers script. The runtime configuration lists the types of server processes to run on an instance, how to launch them, and the number of processes to run concurrently.

Request Lenses

updateRuntimeConfiguration_fleetId :: Lens' UpdateRuntimeConfiguration Text

A unique identifier for the fleet to update runtime configuration for. You can use either the fleet ID or ARN value.

updateRuntimeConfiguration_runtimeConfiguration :: Lens' UpdateRuntimeConfiguration RuntimeConfiguration

Instructions for launching server processes on each instance in the fleet. Server processes run either a custom game build executable or a Realtime Servers script. The runtime configuration lists the types of server processes to run on an instance, how to launch them, and the number of processes to run concurrently.

Destructuring the Response

data UpdateRuntimeConfigurationResponse

Represents the returned data in response to a request operation.

See: newUpdateRuntimeConfigurationResponse smart constructor.

Constructors

UpdateRuntimeConfigurationResponse'

Fields

runtimeConfiguration :: Maybe RuntimeConfiguration

The runtime configuration currently in use by all instances in the fleet. If the update was successful, all property changes are shown.

httpStatus :: Int

The response's http status code.

newUpdateRuntimeConfigurationResponse

Arguments

:: Int
$sel:httpStatus:UpdateRuntimeConfigurationResponse'

-> UpdateRuntimeConfigurationResponse

Create a value of UpdateRuntimeConfigurationResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:runtimeConfiguration:UpdateRuntimeConfiguration', updateRuntimeConfigurationResponse_runtimeConfiguration - The runtime configuration currently in use by all instances in the fleet. If the update was successful, all property changes are shown.

$sel:httpStatus:UpdateRuntimeConfigurationResponse', updateRuntimeConfigurationResponse_httpStatus - The response's http status code.

Response Lenses

updateRuntimeConfigurationResponse_runtimeConfiguration :: Lens' UpdateRuntimeConfigurationResponse (Maybe RuntimeConfiguration)

The runtime configuration currently in use by all instances in the fleet. If the update was successful, all property changes are shown.

updateRuntimeConfigurationResponse_httpStatus :: Lens' UpdateRuntimeConfigurationResponse Int

The response's http status code.