Amazonka.Glacier.AbortVaultLock

Description

This operation aborts the vault locking process if the vault lock is not in the Locked state. If the vault lock is in the Locked state when this operation is requested, the operation returns an AccessDeniedException error. Aborting the vault locking process removes the vault lock policy from the specified vault.

A vault lock is put into the InProgress state by calling InitiateVaultLock. A vault lock is put into the Locked state by calling CompleteVaultLock. You can get the state of a vault lock by calling GetVaultLock. For more information about the vault locking process, see Amazon Glacier Vault Lock. For more information about vault lock policies, see Amazon Glacier Access Control with Vault Lock Policies.

This operation is idempotent. You can successfully invoke this operation multiple times, if the vault lock is in the InProgress state or if there is no policy associated with the vault.

Creating a Request

data AbortVaultLock

The input values for AbortVaultLock.

See: newAbortVaultLock smart constructor.

Constructors

AbortVaultLock'

Fields

accountId :: Text

The AccountId value is the AWS account ID. This value must match the AWS account ID associated with the credentials used to sign the request. You can either specify an AWS account ID or optionally a single '-' (hyphen), in which case Amazon Glacier uses the AWS account ID associated with the credentials used to sign the request. If you specify your account ID, do not include any hyphens ('-') in the ID.

vaultName :: Text

The name of the vault.

newAbortVaultLock

Arguments

:: Text
$sel:accountId:AbortVaultLock'

-> Text
$sel:vaultName:AbortVaultLock'

-> AbortVaultLock

Create a value of AbortVaultLock with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:accountId:AbortVaultLock', abortVaultLock_accountId - The AccountId value is the AWS account ID. This value must match the AWS account ID associated with the credentials used to sign the request. You can either specify an AWS account ID or optionally a single '-' (hyphen), in which case Amazon Glacier uses the AWS account ID associated with the credentials used to sign the request. If you specify your account ID, do not include any hyphens ('-') in the ID.

$sel:vaultName:AbortVaultLock', abortVaultLock_vaultName - The name of the vault.

Request Lenses

abortVaultLock_accountId :: Lens' AbortVaultLock Text

The AccountId value is the AWS account ID. This value must match the AWS account ID associated with the credentials used to sign the request. You can either specify an AWS account ID or optionally a single '-' (hyphen), in which case Amazon Glacier uses the AWS account ID associated with the credentials used to sign the request. If you specify your account ID, do not include any hyphens ('-') in the ID.

abortVaultLock_vaultName :: Lens' AbortVaultLock Text

The name of the vault.

Destructuring the Response

data AbortVaultLockResponse

See: newAbortVaultLockResponse smart constructor.

Constructors

AbortVaultLockResponse'

newAbortVaultLockResponse :: AbortVaultLockResponse

Create a value of AbortVaultLockResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.