Amazonka.Glacier.GetVaultAccessPolicy

Description

This operation retrieves the access-policy subresource set on the vault; for more information on setting this subresource, see Set Vault Access Policy (PUT access-policy). If there is no access policy set on the vault, the operation returns a 404 Not found error. For more information about vault access policies, see Amazon Glacier Access Control with Vault Access Policies.

Creating a Request

data GetVaultAccessPolicy

Input for GetVaultAccessPolicy.

See: newGetVaultAccessPolicy smart constructor.

Constructors

GetVaultAccessPolicy'

Fields

accountId :: Text

The AccountId value is the AWS account ID of the account that owns the vault. You can either specify an AWS account ID or optionally a single '-' (hyphen), in which case Amazon S3 Glacier uses the AWS account ID associated with the credentials used to sign the request. If you use an account ID, do not include any hyphens ('-') in the ID.

vaultName :: Text

The name of the vault.

newGetVaultAccessPolicy

Arguments

:: Text
$sel:accountId:GetVaultAccessPolicy'

-> Text
$sel:vaultName:GetVaultAccessPolicy'

-> GetVaultAccessPolicy

Create a value of GetVaultAccessPolicy with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:accountId:GetVaultAccessPolicy', getVaultAccessPolicy_accountId - The AccountId value is the AWS account ID of the account that owns the vault. You can either specify an AWS account ID or optionally a single '-' (hyphen), in which case Amazon S3 Glacier uses the AWS account ID associated with the credentials used to sign the request. If you use an account ID, do not include any hyphens ('-') in the ID.

$sel:vaultName:GetVaultAccessPolicy', getVaultAccessPolicy_vaultName - The name of the vault.

Request Lenses

getVaultAccessPolicy_accountId :: Lens' GetVaultAccessPolicy Text

The AccountId value is the AWS account ID of the account that owns the vault. You can either specify an AWS account ID or optionally a single '-' (hyphen), in which case Amazon S3 Glacier uses the AWS account ID associated with the credentials used to sign the request. If you use an account ID, do not include any hyphens ('-') in the ID.

getVaultAccessPolicy_vaultName :: Lens' GetVaultAccessPolicy Text

The name of the vault.

Destructuring the Response

data GetVaultAccessPolicyResponse

Output for GetVaultAccessPolicy.

See: newGetVaultAccessPolicyResponse smart constructor.

Constructors

GetVaultAccessPolicyResponse'

Fields

policy :: Maybe VaultAccessPolicy

Contains the returned vault access policy as a JSON string.

httpStatus :: Int

The response's http status code.

newGetVaultAccessPolicyResponse

Arguments

:: Int
$sel:httpStatus:GetVaultAccessPolicyResponse'

-> GetVaultAccessPolicyResponse

Create a value of GetVaultAccessPolicyResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:policy:GetVaultAccessPolicyResponse', getVaultAccessPolicyResponse_policy - Contains the returned vault access policy as a JSON string.

$sel:httpStatus:GetVaultAccessPolicyResponse', getVaultAccessPolicyResponse_httpStatus - The response's http status code.

Response Lenses

getVaultAccessPolicyResponse_policy :: Lens' GetVaultAccessPolicyResponse (Maybe VaultAccessPolicy)

Contains the returned vault access policy as a JSON string.

getVaultAccessPolicyResponse_httpStatus :: Lens' GetVaultAccessPolicyResponse Int

The response's http status code.