Amazonka.Glacier.GetVaultLock

Description

This operation retrieves the following attributes from the lock-policy subresource set on the specified vault:

The vault lock policy set on the vault.

The state of the vault lock, which is either InProgess or Locked.

When the lock ID expires. The lock ID is used to complete the vault locking process.

When the vault lock was initiated and put into the InProgress state.

A vault lock is put into the InProgress state by calling InitiateVaultLock. A vault lock is put into the Locked state by calling CompleteVaultLock. You can abort the vault locking process by calling AbortVaultLock. For more information about the vault locking process, Amazon Glacier Vault Lock.

If there is no vault lock policy set on the vault, the operation returns a 404 Not found error. For more information about vault lock policies, Amazon Glacier Access Control with Vault Lock Policies.

Creating a Request

data GetVaultLock

The input values for GetVaultLock.

See: newGetVaultLock smart constructor.

Constructors

GetVaultLock'

Fields

accountId :: Text

The AccountId value is the AWS account ID of the account that owns the vault. You can either specify an AWS account ID or optionally a single '-' (hyphen), in which case Amazon S3 Glacier uses the AWS account ID associated with the credentials used to sign the request. If you use an account ID, do not include any hyphens ('-') in the ID.

vaultName :: Text

The name of the vault.

newGetVaultLock

Arguments

:: Text
$sel:accountId:GetVaultLock'

-> Text
$sel:vaultName:GetVaultLock'

-> GetVaultLock

Create a value of GetVaultLock with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:accountId:GetVaultLock', getVaultLock_accountId - The AccountId value is the AWS account ID of the account that owns the vault. You can either specify an AWS account ID or optionally a single '-' (hyphen), in which case Amazon S3 Glacier uses the AWS account ID associated with the credentials used to sign the request. If you use an account ID, do not include any hyphens ('-') in the ID.

$sel:vaultName:GetVaultLock', getVaultLock_vaultName - The name of the vault.

Request Lenses

getVaultLock_accountId :: Lens' GetVaultLock Text

The AccountId value is the AWS account ID of the account that owns the vault. You can either specify an AWS account ID or optionally a single '-' (hyphen), in which case Amazon S3 Glacier uses the AWS account ID associated with the credentials used to sign the request. If you use an account ID, do not include any hyphens ('-') in the ID.

getVaultLock_vaultName :: Lens' GetVaultLock Text

The name of the vault.

Destructuring the Response

data GetVaultLockResponse

Contains the Amazon S3 Glacier response to your request.

See: newGetVaultLockResponse smart constructor.

Constructors

GetVaultLockResponse'

Fields

state :: Maybe Text

The state of the vault lock. InProgress or Locked.

expirationDate :: Maybe Text

The UTC date and time at which the lock ID expires. This value can be null if the vault lock is in a Locked state.

creationDate :: Maybe Text

The UTC date and time at which the vault lock was put into the InProgress state.

policy :: Maybe Text

The vault lock policy as a JSON string, which uses "\" as an escape character.

httpStatus :: Int

The response's http status code.

newGetVaultLockResponse

Arguments

:: Int
$sel:httpStatus:GetVaultLockResponse'

-> GetVaultLockResponse

Create a value of GetVaultLockResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:state:GetVaultLockResponse', getVaultLockResponse_state - The state of the vault lock. InProgress or Locked.

$sel:expirationDate:GetVaultLockResponse', getVaultLockResponse_expirationDate - The UTC date and time at which the lock ID expires. This value can be null if the vault lock is in a Locked state.

$sel:creationDate:GetVaultLockResponse', getVaultLockResponse_creationDate - The UTC date and time at which the vault lock was put into the InProgress state.

$sel:policy:GetVaultLockResponse', getVaultLockResponse_policy - The vault lock policy as a JSON string, which uses "\" as an escape character.

$sel:httpStatus:GetVaultLockResponse', getVaultLockResponse_httpStatus - The response's http status code.

Response Lenses

getVaultLockResponse_state :: Lens' GetVaultLockResponse (Maybe Text)

The state of the vault lock. InProgress or Locked.

getVaultLockResponse_expirationDate :: Lens' GetVaultLockResponse (Maybe Text)

The UTC date and time at which the lock ID expires. This value can be null if the vault lock is in a Locked state.

getVaultLockResponse_creationDate :: Lens' GetVaultLockResponse (Maybe Text)

The UTC date and time at which the vault lock was put into the InProgress state.

getVaultLockResponse_policy :: Lens' GetVaultLockResponse (Maybe Text)

The vault lock policy as a JSON string, which uses "\" as an escape character.

getVaultLockResponse_httpStatus :: Lens' GetVaultLockResponse Int

The response's http status code.