Amazonka.Glacier.ListMultipartUploads

Description

This operation lists in-progress multipart uploads for the specified vault. An in-progress multipart upload is a multipart upload that has been initiated by an InitiateMultipartUpload request, but has not yet been completed or aborted. The list returned in the List Multipart Upload response has no guaranteed order.

The List Multipart Uploads operation supports pagination. By default, this operation returns up to 50 multipart uploads in the response. You should always check the response for a marker at which to continue the list; if there are no more items the marker is null. To return a list of multipart uploads that begins at a specific upload, set the marker request parameter to the value you obtained from a previous List Multipart Upload request. You can also limit the number of uploads returned in the response by specifying the limit parameter in the request.

Note the difference between this operation and listing parts (ListParts). The List Multipart Uploads operation lists all multipart uploads for a vault and does not require a multipart upload ID. The List Parts operation requires a multipart upload ID since parts are associated with a single upload.

An AWS account has full permission to perform all operations (actions). However, AWS Identity and Access Management (IAM) users don't have any permissions by default. You must grant them explicit permission to perform specific actions. For more information, see Access Control Using AWS Identity and Access Management (IAM).

For conceptual information and the underlying REST API, see Working with Archives in Amazon S3 Glacier and List Multipart Uploads in the Amazon Glacier Developer Guide.

This operation returns paginated results.

Creating a Request

data ListMultipartUploads

Provides options for retrieving list of in-progress multipart uploads for an Amazon Glacier vault.

See: newListMultipartUploads smart constructor.

Constructors

ListMultipartUploads'

Fields

marker :: Maybe Text

An opaque string used for pagination. This value specifies the upload at which the listing of uploads should begin. Get the marker value from a previous List Uploads response. You need only include the marker if you are continuing the pagination of results started in a previous List Uploads request.

limit :: Maybe Text

Specifies the maximum number of uploads returned in the response body. If this value is not specified, the List Uploads operation returns up to 50 uploads.

accountId :: Text

The AccountId value is the AWS account ID of the account that owns the vault. You can either specify an AWS account ID or optionally a single '-' (hyphen), in which case Amazon S3 Glacier uses the AWS account ID associated with the credentials used to sign the request. If you use an account ID, do not include any hyphens ('-') in the ID.

vaultName :: Text

The name of the vault.

newListMultipartUploads

Arguments

:: Text
$sel:accountId:ListMultipartUploads'

-> Text
$sel:vaultName:ListMultipartUploads'

-> ListMultipartUploads

Create a value of ListMultipartUploads with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:marker:ListMultipartUploads', listMultipartUploads_marker - An opaque string used for pagination. This value specifies the upload at which the listing of uploads should begin. Get the marker value from a previous List Uploads response. You need only include the marker if you are continuing the pagination of results started in a previous List Uploads request.

$sel:limit:ListMultipartUploads', listMultipartUploads_limit - Specifies the maximum number of uploads returned in the response body. If this value is not specified, the List Uploads operation returns up to 50 uploads.

$sel:accountId:ListMultipartUploads', listMultipartUploads_accountId - The AccountId value is the AWS account ID of the account that owns the vault. You can either specify an AWS account ID or optionally a single '-' (hyphen), in which case Amazon S3 Glacier uses the AWS account ID associated with the credentials used to sign the request. If you use an account ID, do not include any hyphens ('-') in the ID.

$sel:vaultName:ListMultipartUploads', listMultipartUploads_vaultName - The name of the vault.

Request Lenses

listMultipartUploads_marker :: Lens' ListMultipartUploads (Maybe Text)

An opaque string used for pagination. This value specifies the upload at which the listing of uploads should begin. Get the marker value from a previous List Uploads response. You need only include the marker if you are continuing the pagination of results started in a previous List Uploads request.

listMultipartUploads_limit :: Lens' ListMultipartUploads (Maybe Text)

Specifies the maximum number of uploads returned in the response body. If this value is not specified, the List Uploads operation returns up to 50 uploads.

listMultipartUploads_accountId :: Lens' ListMultipartUploads Text

The AccountId value is the AWS account ID of the account that owns the vault. You can either specify an AWS account ID or optionally a single '-' (hyphen), in which case Amazon S3 Glacier uses the AWS account ID associated with the credentials used to sign the request. If you use an account ID, do not include any hyphens ('-') in the ID.

listMultipartUploads_vaultName :: Lens' ListMultipartUploads Text

The name of the vault.

Destructuring the Response

data ListMultipartUploadsResponse

Contains the Amazon S3 Glacier response to your request.

See: newListMultipartUploadsResponse smart constructor.

Constructors

ListMultipartUploadsResponse'

Fields

uploadsList :: Maybe [UploadListElement]

A list of in-progress multipart uploads.

marker :: Maybe Text

An opaque string that represents where to continue pagination of the results. You use the marker in a new List Multipart Uploads request to obtain more uploads in the list. If there are no more uploads, this value is null.

httpStatus :: Int

The response's http status code.

newListMultipartUploadsResponse

Arguments

:: Int
$sel:httpStatus:ListMultipartUploadsResponse'

-> ListMultipartUploadsResponse

Create a value of ListMultipartUploadsResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:uploadsList:ListMultipartUploadsResponse', listMultipartUploadsResponse_uploadsList - A list of in-progress multipart uploads.

$sel:marker:ListMultipartUploads', listMultipartUploadsResponse_marker - An opaque string that represents where to continue pagination of the results. You use the marker in a new List Multipart Uploads request to obtain more uploads in the list. If there are no more uploads, this value is null.

$sel:httpStatus:ListMultipartUploadsResponse', listMultipartUploadsResponse_httpStatus - The response's http status code.

Response Lenses

listMultipartUploadsResponse_uploadsList :: Lens' ListMultipartUploadsResponse (Maybe [UploadListElement])

A list of in-progress multipart uploads.

listMultipartUploadsResponse_marker :: Lens' ListMultipartUploadsResponse (Maybe Text)

An opaque string that represents where to continue pagination of the results. You use the marker in a new List Multipart Uploads request to obtain more uploads in the list. If there are no more uploads, this value is null.

listMultipartUploadsResponse_httpStatus :: Lens' ListMultipartUploadsResponse Int

The response's http status code.