Amazonka.Glacier.ListParts

Description

This operation lists the parts of an archive that have been uploaded in a specific multipart upload. You can make this request at any time during an in-progress multipart upload before you complete the upload (see CompleteMultipartUpload. List Parts returns an error for completed uploads. The list returned in the List Parts response is sorted by part range.

The List Parts operation supports pagination. By default, this operation returns up to 50 uploaded parts in the response. You should always check the response for a marker at which to continue the list; if there are no more items the marker is null. To return a list of parts that begins at a specific part, set the marker request parameter to the value you obtained from a previous List Parts request. You can also limit the number of parts returned in the response by specifying the limit parameter in the request.

An AWS account has full permission to perform all operations (actions). However, AWS Identity and Access Management (IAM) users don't have any permissions by default. You must grant them explicit permission to perform specific actions. For more information, see Access Control Using AWS Identity and Access Management (IAM).

For conceptual information and the underlying REST API, see Working with Archives in Amazon S3 Glacier and List Parts in the Amazon Glacier Developer Guide.

This operation returns paginated results.

Creating a Request

data ListParts

Provides options for retrieving a list of parts of an archive that have been uploaded in a specific multipart upload.

See: newListParts smart constructor.

Constructors

ListParts'

Fields

marker :: Maybe Text

An opaque string used for pagination. This value specifies the part at which the listing of parts should begin. Get the marker value from the response of a previous List Parts response. You need only include the marker if you are continuing the pagination of results started in a previous List Parts request.

limit :: Maybe Text

The maximum number of parts to be returned. The default limit is 50. The number of parts returned might be fewer than the specified limit, but the number of returned parts never exceeds the limit.

accountId :: Text

The AccountId value is the AWS account ID of the account that owns the vault. You can either specify an AWS account ID or optionally a single '-' (hyphen), in which case Amazon S3 Glacier uses the AWS account ID associated with the credentials used to sign the request. If you use an account ID, do not include any hyphens ('-') in the ID.

vaultName :: Text

The name of the vault.

uploadId :: Text

The upload ID of the multipart upload.

newListParts

Arguments

:: Text
$sel:accountId:ListParts'

-> Text
$sel:vaultName:ListParts'

-> Text
$sel:uploadId:ListParts'

-> ListParts

Create a value of ListParts with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:marker:ListParts', listParts_marker - An opaque string used for pagination. This value specifies the part at which the listing of parts should begin. Get the marker value from the response of a previous List Parts response. You need only include the marker if you are continuing the pagination of results started in a previous List Parts request.

$sel:limit:ListParts', listParts_limit - The maximum number of parts to be returned. The default limit is 50. The number of parts returned might be fewer than the specified limit, but the number of returned parts never exceeds the limit.

$sel:accountId:ListParts', listParts_accountId - The AccountId value is the AWS account ID of the account that owns the vault. You can either specify an AWS account ID or optionally a single '-' (hyphen), in which case Amazon S3 Glacier uses the AWS account ID associated with the credentials used to sign the request. If you use an account ID, do not include any hyphens ('-') in the ID.

$sel:vaultName:ListParts', listParts_vaultName - The name of the vault.

$sel:uploadId:ListParts', listParts_uploadId - The upload ID of the multipart upload.

Request Lenses

listParts_marker :: Lens' ListParts (Maybe Text)

An opaque string used for pagination. This value specifies the part at which the listing of parts should begin. Get the marker value from the response of a previous List Parts response. You need only include the marker if you are continuing the pagination of results started in a previous List Parts request.

listParts_limit :: Lens' ListParts (Maybe Text)

The maximum number of parts to be returned. The default limit is 50. The number of parts returned might be fewer than the specified limit, but the number of returned parts never exceeds the limit.

listParts_accountId :: Lens' ListParts Text

The AccountId value is the AWS account ID of the account that owns the vault. You can either specify an AWS account ID or optionally a single '-' (hyphen), in which case Amazon S3 Glacier uses the AWS account ID associated with the credentials used to sign the request. If you use an account ID, do not include any hyphens ('-') in the ID.

listParts_vaultName :: Lens' ListParts Text

The name of the vault.

listParts_uploadId :: Lens' ListParts Text

The upload ID of the multipart upload.

Destructuring the Response

data ListPartsResponse

Contains the Amazon S3 Glacier response to your request.

See: newListPartsResponse smart constructor.

Constructors

ListPartsResponse'

Fields

parts :: Maybe [PartListElement]

A list of the part sizes of the multipart upload. Each object in the array contains a RangeBytes and sha256-tree-hash name/value pair.

multipartUploadId :: Maybe Text

The ID of the upload to which the parts are associated.

partSizeInBytes :: Maybe Integer

The part size in bytes. This is the same value that you specified in the Initiate Multipart Upload request.

archiveDescription :: Maybe Text

The description of the archive that was specified in the Initiate Multipart Upload request.

vaultARN :: Maybe Text

The Amazon Resource Name (ARN) of the vault to which the multipart upload was initiated.

marker :: Maybe Text

An opaque string that represents where to continue pagination of the results. You use the marker in a new List Parts request to obtain more jobs in the list. If there are no more parts, this value is null.

creationDate :: Maybe Text

The UTC time at which the multipart upload was initiated.

httpStatus :: Int

The response's http status code.

newListPartsResponse

Arguments

:: Int
$sel:httpStatus:ListPartsResponse'

-> ListPartsResponse

Create a value of ListPartsResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:parts:ListPartsResponse', listPartsResponse_parts - A list of the part sizes of the multipart upload. Each object in the array contains a RangeBytes and sha256-tree-hash name/value pair.

$sel:multipartUploadId:ListPartsResponse', listPartsResponse_multipartUploadId - The ID of the upload to which the parts are associated.

$sel:partSizeInBytes:ListPartsResponse', listPartsResponse_partSizeInBytes - The part size in bytes. This is the same value that you specified in the Initiate Multipart Upload request.

$sel:archiveDescription:ListPartsResponse', listPartsResponse_archiveDescription - The description of the archive that was specified in the Initiate Multipart Upload request.

$sel:vaultARN:ListPartsResponse', listPartsResponse_vaultARN - The Amazon Resource Name (ARN) of the vault to which the multipart upload was initiated.

$sel:marker:ListParts', listPartsResponse_marker - An opaque string that represents where to continue pagination of the results. You use the marker in a new List Parts request to obtain more jobs in the list. If there are no more parts, this value is null.

$sel:creationDate:ListPartsResponse', listPartsResponse_creationDate - The UTC time at which the multipart upload was initiated.

$sel:httpStatus:ListPartsResponse', listPartsResponse_httpStatus - The response's http status code.

Response Lenses

listPartsResponse_parts :: Lens' ListPartsResponse (Maybe [PartListElement])

A list of the part sizes of the multipart upload. Each object in the array contains a RangeBytes and sha256-tree-hash name/value pair.

listPartsResponse_multipartUploadId :: Lens' ListPartsResponse (Maybe Text)

The ID of the upload to which the parts are associated.

listPartsResponse_partSizeInBytes :: Lens' ListPartsResponse (Maybe Integer)

The part size in bytes. This is the same value that you specified in the Initiate Multipart Upload request.

listPartsResponse_archiveDescription :: Lens' ListPartsResponse (Maybe Text)

The description of the archive that was specified in the Initiate Multipart Upload request.

listPartsResponse_vaultARN :: Lens' ListPartsResponse (Maybe Text)

The Amazon Resource Name (ARN) of the vault to which the multipart upload was initiated.

listPartsResponse_marker :: Lens' ListPartsResponse (Maybe Text)

An opaque string that represents where to continue pagination of the results. You use the marker in a new List Parts request to obtain more jobs in the list. If there are no more parts, this value is null.

listPartsResponse_creationDate :: Lens' ListPartsResponse (Maybe Text)

The UTC time at which the multipart upload was initiated.

listPartsResponse_httpStatus :: Lens' ListPartsResponse Int

The response's http status code.