Amazonka.Glacier.SetDataRetrievalPolicy

Description

This operation sets and then enacts a data retrieval policy in the region specified in the PUT request. You can set one policy per region for an AWS account. The policy is enacted within a few minutes of a successful PUT operation.

The set policy operation does not affect retrieval jobs that were in progress before the policy was enacted. For more information about data retrieval policies, see Amazon Glacier Data Retrieval Policies.

Creating a Request

data SetDataRetrievalPolicy

SetDataRetrievalPolicy input.

See: newSetDataRetrievalPolicy smart constructor.

Constructors

SetDataRetrievalPolicy'

Fields

policy :: Maybe DataRetrievalPolicy

The data retrieval policy in JSON format.

accountId :: Text

The AccountId value is the AWS account ID. This value must match the AWS account ID associated with the credentials used to sign the request. You can either specify an AWS account ID or optionally a single '-' (hyphen), in which case Amazon Glacier uses the AWS account ID associated with the credentials used to sign the request. If you specify your account ID, do not include any hyphens ('-') in the ID.

newSetDataRetrievalPolicy

Arguments

:: Text
$sel:accountId:SetDataRetrievalPolicy'

-> SetDataRetrievalPolicy

Create a value of SetDataRetrievalPolicy with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:policy:SetDataRetrievalPolicy', setDataRetrievalPolicy_policy - The data retrieval policy in JSON format.

$sel:accountId:SetDataRetrievalPolicy', setDataRetrievalPolicy_accountId - The AccountId value is the AWS account ID. This value must match the AWS account ID associated with the credentials used to sign the request. You can either specify an AWS account ID or optionally a single '-' (hyphen), in which case Amazon Glacier uses the AWS account ID associated with the credentials used to sign the request. If you specify your account ID, do not include any hyphens ('-') in the ID.

Request Lenses

setDataRetrievalPolicy_policy :: Lens' SetDataRetrievalPolicy (Maybe DataRetrievalPolicy)

The data retrieval policy in JSON format.

setDataRetrievalPolicy_accountId :: Lens' SetDataRetrievalPolicy Text

The AccountId value is the AWS account ID. This value must match the AWS account ID associated with the credentials used to sign the request. You can either specify an AWS account ID or optionally a single '-' (hyphen), in which case Amazon Glacier uses the AWS account ID associated with the credentials used to sign the request. If you specify your account ID, do not include any hyphens ('-') in the ID.

Destructuring the Response

data SetDataRetrievalPolicyResponse

See: newSetDataRetrievalPolicyResponse smart constructor.

Constructors

SetDataRetrievalPolicyResponse'

newSetDataRetrievalPolicyResponse :: SetDataRetrievalPolicyResponse

Create a value of SetDataRetrievalPolicyResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.