Amazonka.Glacier.Types.GlacierJobDescription

Description

Documentation

data GlacierJobDescription

Contains the description of an Amazon S3 Glacier job.

See: newGlacierJobDescription smart constructor.

Constructors

GlacierJobDescription'

Fields

sHA256TreeHash :: Maybe Text

For an archive retrieval job, this value is the checksum of the archive. Otherwise, this value is null.

The SHA256 tree hash value for the requested range of an archive. If the InitiateJob request for an archive specified a tree-hash aligned range, then this field returns a value.

If the whole archive is retrieved, this value is the same as the ArchiveSHA256TreeHash value.

This field is null for the following:

Archive retrieval jobs that specify a range that is not tree-hash aligned

Archival jobs that specify a range that is equal to the whole archive, when the job status is InProgress

Inventory jobs

Select jobs

archiveId :: Maybe Text

The archive ID requested for a select job or archive retrieval. Otherwise, this field is null.

selectParameters :: Maybe SelectParameters

Contains the parameters used for a select.

jobId :: Maybe Text

An opaque string that identifies an Amazon S3 Glacier job.

jobOutputPath :: Maybe Text

Contains the job output location.

retrievalByteRange :: Maybe Text

The retrieved byte range for archive retrieval jobs in the form StartByteValue-EndByteValue. If no range was specified in the archive retrieval, then the whole archive is retrieved. In this case, StartByteValue equals 0 and EndByteValue equals the size of the archive minus 1. For inventory retrieval or select jobs, this field is null.

inventoryRetrievalParameters :: Maybe InventoryRetrievalJobDescription

Parameters used for range inventory retrieval.

action :: Maybe ActionCode

The job type. This value is either ArchiveRetrieval, InventoryRetrieval, or Select.

jobDescription :: Maybe Text

The job description provided when initiating the job.

sNSTopic :: Maybe Text

An Amazon SNS topic that receives notification.

statusMessage :: Maybe Text

A friendly message that describes the job status.

vaultARN :: Maybe Text

The Amazon Resource Name (ARN) of the vault from which an archive retrieval was requested.

outputLocation :: Maybe OutputLocation

Contains the location where the data from the select job is stored.

tier :: Maybe Text

The tier to use for a select or an archive retrieval. Valid values are Expedited, Standard, or Bulk. Standard is the default.

archiveSHA256TreeHash :: Maybe Text

The SHA256 tree hash of the entire archive for an archive retrieval. For inventory retrieval or select jobs, this field is null.

creationDate :: Maybe Text

The UTC date when the job was created. This value is a string representation of ISO 8601 date format, for example "2012-03-20T17:03:43.221Z".

completed :: Maybe Bool

The job status. When a job is completed, you get the job's output using Get Job Output (GET output).

completionDate :: Maybe Text

The UTC time that the job request completed. While the job is in progress, the value is null.

inventorySizeInBytes :: Maybe Integer

For an inventory retrieval job, this value is the size in bytes of the inventory requested for download. For an archive retrieval or select job, this value is null.

archiveSizeInBytes :: Maybe Integer

For an archive retrieval job, this value is the size in bytes of the archive being requested for download. For an inventory retrieval or select job, this value is null.

statusCode :: Maybe StatusCode

The status code can be InProgress, Succeeded, or Failed, and indicates the status of the job.

newGlacierJobDescription :: GlacierJobDescription

Create a value of GlacierJobDescription with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:sHA256TreeHash:GlacierJobDescription', glacierJobDescription_sHA256TreeHash - For an archive retrieval job, this value is the checksum of the archive. Otherwise, this value is null.

The SHA256 tree hash value for the requested range of an archive. If the InitiateJob request for an archive specified a tree-hash aligned range, then this field returns a value.

If the whole archive is retrieved, this value is the same as the ArchiveSHA256TreeHash value.

This field is null for the following:

Archive retrieval jobs that specify a range that is not tree-hash aligned

Archival jobs that specify a range that is equal to the whole archive, when the job status is InProgress

Inventory jobs

Select jobs

$sel:archiveId:GlacierJobDescription', glacierJobDescription_archiveId - The archive ID requested for a select job or archive retrieval. Otherwise, this field is null.

$sel:selectParameters:GlacierJobDescription', glacierJobDescription_selectParameters - Contains the parameters used for a select.

$sel:jobId:GlacierJobDescription', glacierJobDescription_jobId - An opaque string that identifies an Amazon S3 Glacier job.

$sel:jobOutputPath:GlacierJobDescription', glacierJobDescription_jobOutputPath - Contains the job output location.

$sel:retrievalByteRange:GlacierJobDescription', glacierJobDescription_retrievalByteRange - The retrieved byte range for archive retrieval jobs in the form StartByteValue-EndByteValue. If no range was specified in the archive retrieval, then the whole archive is retrieved. In this case, StartByteValue equals 0 and EndByteValue equals the size of the archive minus 1. For inventory retrieval or select jobs, this field is null.

$sel:inventoryRetrievalParameters:GlacierJobDescription', glacierJobDescription_inventoryRetrievalParameters - Parameters used for range inventory retrieval.

$sel:action:GlacierJobDescription', glacierJobDescription_action - The job type. This value is either ArchiveRetrieval, InventoryRetrieval, or Select.

$sel:jobDescription:GlacierJobDescription', glacierJobDescription_jobDescription - The job description provided when initiating the job.

$sel:sNSTopic:GlacierJobDescription', glacierJobDescription_sNSTopic - An Amazon SNS topic that receives notification.

$sel:statusMessage:GlacierJobDescription', glacierJobDescription_statusMessage - A friendly message that describes the job status.

$sel:vaultARN:GlacierJobDescription', glacierJobDescription_vaultARN - The Amazon Resource Name (ARN) of the vault from which an archive retrieval was requested.

$sel:outputLocation:GlacierJobDescription', glacierJobDescription_outputLocation - Contains the location where the data from the select job is stored.

$sel:tier:GlacierJobDescription', glacierJobDescription_tier - The tier to use for a select or an archive retrieval. Valid values are Expedited, Standard, or Bulk. Standard is the default.

$sel:archiveSHA256TreeHash:GlacierJobDescription', glacierJobDescription_archiveSHA256TreeHash - The SHA256 tree hash of the entire archive for an archive retrieval. For inventory retrieval or select jobs, this field is null.

$sel:creationDate:GlacierJobDescription', glacierJobDescription_creationDate - The UTC date when the job was created. This value is a string representation of ISO 8601 date format, for example "2012-03-20T17:03:43.221Z".

$sel:completed:GlacierJobDescription', glacierJobDescription_completed - The job status. When a job is completed, you get the job's output using Get Job Output (GET output).

$sel:completionDate:GlacierJobDescription', glacierJobDescription_completionDate - The UTC time that the job request completed. While the job is in progress, the value is null.

$sel:inventorySizeInBytes:GlacierJobDescription', glacierJobDescription_inventorySizeInBytes - For an inventory retrieval job, this value is the size in bytes of the inventory requested for download. For an archive retrieval or select job, this value is null.

$sel:archiveSizeInBytes:GlacierJobDescription', glacierJobDescription_archiveSizeInBytes - For an archive retrieval job, this value is the size in bytes of the archive being requested for download. For an inventory retrieval or select job, this value is null.

$sel:statusCode:GlacierJobDescription', glacierJobDescription_statusCode - The status code can be InProgress, Succeeded, or Failed, and indicates the status of the job.

glacierJobDescription_sHA256TreeHash :: Lens' GlacierJobDescription (Maybe Text)

For an archive retrieval job, this value is the checksum of the archive. Otherwise, this value is null.

The SHA256 tree hash value for the requested range of an archive. If the InitiateJob request for an archive specified a tree-hash aligned range, then this field returns a value.

If the whole archive is retrieved, this value is the same as the ArchiveSHA256TreeHash value.

This field is null for the following:

Archive retrieval jobs that specify a range that is not tree-hash aligned

Archival jobs that specify a range that is equal to the whole archive, when the job status is InProgress

Inventory jobs

Select jobs

glacierJobDescription_archiveId :: Lens' GlacierJobDescription (Maybe Text)

The archive ID requested for a select job or archive retrieval. Otherwise, this field is null.

glacierJobDescription_selectParameters :: Lens' GlacierJobDescription (Maybe SelectParameters)

Contains the parameters used for a select.

glacierJobDescription_jobId :: Lens' GlacierJobDescription (Maybe Text)

An opaque string that identifies an Amazon S3 Glacier job.

glacierJobDescription_jobOutputPath :: Lens' GlacierJobDescription (Maybe Text)

Contains the job output location.

glacierJobDescription_retrievalByteRange :: Lens' GlacierJobDescription (Maybe Text)

The retrieved byte range for archive retrieval jobs in the form StartByteValue-EndByteValue. If no range was specified in the archive retrieval, then the whole archive is retrieved. In this case, StartByteValue equals 0 and EndByteValue equals the size of the archive minus 1. For inventory retrieval or select jobs, this field is null.

glacierJobDescription_inventoryRetrievalParameters :: Lens' GlacierJobDescription (Maybe InventoryRetrievalJobDescription)

Parameters used for range inventory retrieval.

glacierJobDescription_action :: Lens' GlacierJobDescription (Maybe ActionCode)

The job type. This value is either ArchiveRetrieval, InventoryRetrieval, or Select.

glacierJobDescription_jobDescription :: Lens' GlacierJobDescription (Maybe Text)

The job description provided when initiating the job.

glacierJobDescription_sNSTopic :: Lens' GlacierJobDescription (Maybe Text)

An Amazon SNS topic that receives notification.

glacierJobDescription_statusMessage :: Lens' GlacierJobDescription (Maybe Text)

A friendly message that describes the job status.

glacierJobDescription_vaultARN :: Lens' GlacierJobDescription (Maybe Text)

The Amazon Resource Name (ARN) of the vault from which an archive retrieval was requested.

glacierJobDescription_outputLocation :: Lens' GlacierJobDescription (Maybe OutputLocation)

Contains the location where the data from the select job is stored.

glacierJobDescription_tier :: Lens' GlacierJobDescription (Maybe Text)

The tier to use for a select or an archive retrieval. Valid values are Expedited, Standard, or Bulk. Standard is the default.

glacierJobDescription_archiveSHA256TreeHash :: Lens' GlacierJobDescription (Maybe Text)

The SHA256 tree hash of the entire archive for an archive retrieval. For inventory retrieval or select jobs, this field is null.

glacierJobDescription_creationDate :: Lens' GlacierJobDescription (Maybe Text)

The UTC date when the job was created. This value is a string representation of ISO 8601 date format, for example "2012-03-20T17:03:43.221Z".

glacierJobDescription_completed :: Lens' GlacierJobDescription (Maybe Bool)

The job status. When a job is completed, you get the job's output using Get Job Output (GET output).

glacierJobDescription_completionDate :: Lens' GlacierJobDescription (Maybe Text)

The UTC time that the job request completed. While the job is in progress, the value is null.

glacierJobDescription_inventorySizeInBytes :: Lens' GlacierJobDescription (Maybe Integer)

For an inventory retrieval job, this value is the size in bytes of the inventory requested for download. For an archive retrieval or select job, this value is null.

glacierJobDescription_archiveSizeInBytes :: Lens' GlacierJobDescription (Maybe Integer)

For an archive retrieval job, this value is the size in bytes of the archive being requested for download. For an inventory retrieval or select job, this value is null.

glacierJobDescription_statusCode :: Lens' GlacierJobDescription (Maybe StatusCode)

The status code can be InProgress, Succeeded, or Failed, and indicates the status of the job.