Amazonka.Glacier.UploadArchive

Description

This operation adds an archive to a vault. This is a synchronous operation, and for a successful upload, your data is durably persisted. Amazon S3 Glacier returns the archive ID in the x-amz-archive-id header of the response.

You must use the archive ID to access your data in Amazon S3 Glacier. After you upload an archive, you should save the archive ID returned so that you can retrieve or delete the archive later. Besides saving the archive ID, you can also index it and give it a friendly name to allow for better searching. You can also use the optional archive description field to specify how the archive is referred to in an external index of archives, such as you might create in Amazon DynamoDB. You can also get the vault inventory to obtain a list of archive IDs in a vault. For more information, see InitiateJob.

You must provide a SHA256 tree hash of the data you are uploading. For information about computing a SHA256 tree hash, see Computing Checksums.

You can optionally specify an archive description of up to 1,024 printable ASCII characters. You can get the archive description when you either retrieve the archive or get the vault inventory. For more information, see InitiateJob. Amazon Glacier does not interpret the description in any way. An archive description does not need to be unique. You cannot use the description to retrieve or sort the archive list.

Archives are immutable. After you upload an archive, you cannot edit the archive or its description.

An AWS account has full permission to perform all operations (actions). However, AWS Identity and Access Management (IAM) users don't have any permissions by default. You must grant them explicit permission to perform specific actions. For more information, see Access Control Using AWS Identity and Access Management (IAM).

For conceptual information and underlying REST API, see Uploading an Archive in Amazon Glacier and Upload Archive in the Amazon Glacier Developer Guide.

Creating a Request

data UploadArchive

Provides options to add an archive to a vault.

See: newUploadArchive smart constructor.

Constructors

UploadArchive'

Fields

checksum :: Maybe Text

The SHA256 tree hash of the data being uploaded.

archiveDescription :: Maybe Text

The optional description of the archive you are uploading.

vaultName :: Text

The name of the vault.

accountId :: Text

The AccountId value is the AWS account ID of the account that owns the vault. You can either specify an AWS account ID or optionally a single '-' (hyphen), in which case Amazon S3 Glacier uses the AWS account ID associated with the credentials used to sign the request. If you use an account ID, do not include any hyphens ('-') in the ID.

body :: HashedBody

The data to upload.

newUploadArchive

Arguments

:: Text
$sel:vaultName:UploadArchive'

-> Text
$sel:accountId:UploadArchive'

-> HashedBody
$sel:body:UploadArchive'

-> UploadArchive

Create a value of UploadArchive with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:checksum:UploadArchive', uploadArchive_checksum - The SHA256 tree hash of the data being uploaded.

$sel:archiveDescription:UploadArchive', uploadArchive_archiveDescription - The optional description of the archive you are uploading.

$sel:vaultName:UploadArchive', uploadArchive_vaultName - The name of the vault.

$sel:accountId:UploadArchive', uploadArchive_accountId - The AccountId value is the AWS account ID of the account that owns the vault. You can either specify an AWS account ID or optionally a single '-' (hyphen), in which case Amazon S3 Glacier uses the AWS account ID associated with the credentials used to sign the request. If you use an account ID, do not include any hyphens ('-') in the ID.

$sel:body:UploadArchive', uploadArchive_body - The data to upload.

Request Lenses

uploadArchive_checksum :: Lens' UploadArchive (Maybe Text)

The SHA256 tree hash of the data being uploaded.

uploadArchive_archiveDescription :: Lens' UploadArchive (Maybe Text)

The optional description of the archive you are uploading.

uploadArchive_vaultName :: Lens' UploadArchive Text

The name of the vault.

uploadArchive_accountId :: Lens' UploadArchive Text

The AccountId value is the AWS account ID of the account that owns the vault. You can either specify an AWS account ID or optionally a single '-' (hyphen), in which case Amazon S3 Glacier uses the AWS account ID associated with the credentials used to sign the request. If you use an account ID, do not include any hyphens ('-') in the ID.

uploadArchive_body :: Lens' UploadArchive HashedBody

The data to upload.

Destructuring the Response

data ArchiveCreationOutput

Contains the Amazon S3 Glacier response to your request.

For information about the underlying REST API, see Upload Archive. For conceptual information, see Working with Archives in Amazon S3 Glacier.

See: newArchiveCreationOutput smart constructor.

Constructors

ArchiveCreationOutput'

Fields

archiveId :: Maybe Text

The ID of the archive. This value is also included as part of the location.

checksum :: Maybe Text

The checksum of the archive computed by Amazon S3 Glacier.

location :: Maybe Text

The relative URI path of the newly added archive resource.

newArchiveCreationOutput :: ArchiveCreationOutput

Create a value of ArchiveCreationOutput with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:archiveId:ArchiveCreationOutput', archiveCreationOutput_archiveId - The ID of the archive. This value is also included as part of the location.

$sel:checksum:ArchiveCreationOutput', archiveCreationOutput_checksum - The checksum of the archive computed by Amazon S3 Glacier.

$sel:location:ArchiveCreationOutput', archiveCreationOutput_location - The relative URI path of the newly added archive resource.

Response Lenses

archiveCreationOutput_archiveId :: Lens' ArchiveCreationOutput (Maybe Text)

The ID of the archive. This value is also included as part of the location.

archiveCreationOutput_checksum :: Lens' ArchiveCreationOutput (Maybe Text)

The checksum of the archive computed by Amazon S3 Glacier.

archiveCreationOutput_location :: Lens' ArchiveCreationOutput (Maybe Text)

The relative URI path of the newly added archive resource.