Amazonka.GlobalAccelerator.CreateAccelerator

Description

Create an accelerator. An accelerator includes one or more listeners that process inbound connections and direct traffic to one or more endpoint groups, each of which includes endpoints, such as Network Load Balancers.

Global Accelerator is a global service that supports endpoints in multiple AWS Regions but you must specify the US West (Oregon) Region to create or update accelerators.

Creating a Request

data CreateAccelerator

See: newCreateAccelerator smart constructor.

Constructors

CreateAccelerator'

Fields

enabled :: Maybe Bool

Indicates whether an accelerator is enabled. The value is true or false. The default value is true.

If the value is set to true, an accelerator cannot be deleted. If set to false, the accelerator can be deleted.

ipAddressType :: Maybe IpAddressType

The value for the address type must be IPv4.

tags :: Maybe [Tag]

Create tags for an accelerator.

For more information, see Tagging in AWS Global Accelerator in the AWS Global Accelerator Developer Guide.

ipAddresses :: Maybe [Text]

Optionally, if you've added your own IP address pool to Global Accelerator (BYOIP), you can choose IP addresses from your own pool to use for the accelerator's static IP addresses when you create an accelerator. You can specify one or two addresses, separated by a space. Do not include the /32 suffix.

Only one IP address from each of your IP address ranges can be used for each accelerator. If you specify only one IP address from your IP address range, Global Accelerator assigns a second static IP address for the accelerator from the AWS IP address pool.

Note that you can't update IP addresses for an existing accelerator. To change them, you must create a new accelerator with the new addresses.

For more information, see Bring Your Own IP Addresses (BYOIP) in the AWS Global Accelerator Developer Guide.

name :: Text

The name of an accelerator. The name can have a maximum of 32 characters, must contain only alphanumeric characters or hyphens (-), and must not begin or end with a hyphen.

idempotencyToken :: Text

A unique, case-sensitive identifier that you provide to ensure the idempotency—that is, the uniqueness—of an accelerator.

newCreateAccelerator

Arguments

:: Text
$sel:name:CreateAccelerator'

-> Text
$sel:idempotencyToken:CreateAccelerator'

-> CreateAccelerator

Create a value of CreateAccelerator with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:enabled:CreateAccelerator', createAccelerator_enabled - Indicates whether an accelerator is enabled. The value is true or false. The default value is true.

If the value is set to true, an accelerator cannot be deleted. If set to false, the accelerator can be deleted.

$sel:ipAddressType:CreateAccelerator', createAccelerator_ipAddressType - The value for the address type must be IPv4.

$sel:tags:CreateAccelerator', createAccelerator_tags - Create tags for an accelerator.

For more information, see Tagging in AWS Global Accelerator in the AWS Global Accelerator Developer Guide.

$sel:ipAddresses:CreateAccelerator', createAccelerator_ipAddresses - Optionally, if you've added your own IP address pool to Global Accelerator (BYOIP), you can choose IP addresses from your own pool to use for the accelerator's static IP addresses when you create an accelerator. You can specify one or two addresses, separated by a space. Do not include the /32 suffix.

Only one IP address from each of your IP address ranges can be used for each accelerator. If you specify only one IP address from your IP address range, Global Accelerator assigns a second static IP address for the accelerator from the AWS IP address pool.

Note that you can't update IP addresses for an existing accelerator. To change them, you must create a new accelerator with the new addresses.

For more information, see Bring Your Own IP Addresses (BYOIP) in the AWS Global Accelerator Developer Guide.

$sel:name:CreateAccelerator', createAccelerator_name - The name of an accelerator. The name can have a maximum of 32 characters, must contain only alphanumeric characters or hyphens (-), and must not begin or end with a hyphen.

$sel:idempotencyToken:CreateAccelerator', createAccelerator_idempotencyToken - A unique, case-sensitive identifier that you provide to ensure the idempotency—that is, the uniqueness—of an accelerator.

Request Lenses

createAccelerator_enabled :: Lens' CreateAccelerator (Maybe Bool)

Indicates whether an accelerator is enabled. The value is true or false. The default value is true.

If the value is set to true, an accelerator cannot be deleted. If set to false, the accelerator can be deleted.

createAccelerator_ipAddressType :: Lens' CreateAccelerator (Maybe IpAddressType)

The value for the address type must be IPv4.

createAccelerator_tags :: Lens' CreateAccelerator (Maybe [Tag])

Create tags for an accelerator.

For more information, see Tagging in AWS Global Accelerator in the AWS Global Accelerator Developer Guide.

createAccelerator_ipAddresses :: Lens' CreateAccelerator (Maybe [Text])

Optionally, if you've added your own IP address pool to Global Accelerator (BYOIP), you can choose IP addresses from your own pool to use for the accelerator's static IP addresses when you create an accelerator. You can specify one or two addresses, separated by a space. Do not include the /32 suffix.

Only one IP address from each of your IP address ranges can be used for each accelerator. If you specify only one IP address from your IP address range, Global Accelerator assigns a second static IP address for the accelerator from the AWS IP address pool.

Note that you can't update IP addresses for an existing accelerator. To change them, you must create a new accelerator with the new addresses.

For more information, see Bring Your Own IP Addresses (BYOIP) in the AWS Global Accelerator Developer Guide.

createAccelerator_name :: Lens' CreateAccelerator Text

The name of an accelerator. The name can have a maximum of 32 characters, must contain only alphanumeric characters or hyphens (-), and must not begin or end with a hyphen.

createAccelerator_idempotencyToken :: Lens' CreateAccelerator Text

A unique, case-sensitive identifier that you provide to ensure the idempotency—that is, the uniqueness—of an accelerator.

Destructuring the Response

data CreateAcceleratorResponse

See: newCreateAcceleratorResponse smart constructor.

Constructors

CreateAcceleratorResponse'

Fields

accelerator :: Maybe Accelerator

The accelerator that is created by specifying a listener and the supported IP address types.

httpStatus :: Int

The response's http status code.

newCreateAcceleratorResponse

Arguments

:: Int
$sel:httpStatus:CreateAcceleratorResponse'

-> CreateAcceleratorResponse

Create a value of CreateAcceleratorResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:accelerator:CreateAcceleratorResponse', createAcceleratorResponse_accelerator - The accelerator that is created by specifying a listener and the supported IP address types.

$sel:httpStatus:CreateAcceleratorResponse', createAcceleratorResponse_httpStatus - The response's http status code.

Response Lenses

createAcceleratorResponse_accelerator :: Lens' CreateAcceleratorResponse (Maybe Accelerator)

The accelerator that is created by specifying a listener and the supported IP address types.

createAcceleratorResponse_httpStatus :: Lens' CreateAcceleratorResponse Int

The response's http status code.