Amazonka.GlobalAccelerator.CreateCustomRoutingAccelerator

Description

Create a custom routing accelerator. A custom routing accelerator directs traffic to one of possibly thousands of Amazon EC2 instance destinations running in a single or multiple virtual private clouds (VPC) subnet endpoints.

Be aware that, by default, all destination EC2 instances in a VPC subnet endpoint cannot receive traffic. To enable all destinations to receive traffic, or to specify individual port mappings that can receive traffic, see the AllowCustomRoutingTraffic operation.

Global Accelerator is a global service that supports endpoints in multiple AWS Regions but you must specify the US West (Oregon) Region to create or update accelerators.

Creating a Request

data CreateCustomRoutingAccelerator

See: newCreateCustomRoutingAccelerator smart constructor.

Constructors

CreateCustomRoutingAccelerator'

Fields

enabled :: Maybe Bool

Indicates whether an accelerator is enabled. The value is true or false. The default value is true.

If the value is set to true, an accelerator cannot be deleted. If set to false, the accelerator can be deleted.

ipAddressType :: Maybe IpAddressType

The value for the address type must be IPv4.

tags :: Maybe [Tag]

Create tags for an accelerator.

For more information, see Tagging in AWS Global Accelerator in the AWS Global Accelerator Developer Guide.

ipAddresses :: Maybe [Text]

Optionally, if you've added your own IP address pool to Global Accelerator (BYOIP), you can choose IP addresses from your own pool to use for the accelerator's static IP addresses when you create an accelerator. You can specify one or two addresses, separated by a space. Do not include the /32 suffix.

Only one IP address from each of your IP address ranges can be used for each accelerator. If you specify only one IP address from your IP address range, Global Accelerator assigns a second static IP address for the accelerator from the AWS IP address pool.

Note that you can't update IP addresses for an existing accelerator. To change them, you must create a new accelerator with the new addresses.

For more information, see Bring your own IP addresses (BYOIP) in the AWS Global Accelerator Developer Guide.

name :: Text

The name of a custom routing accelerator. The name can have a maximum of 64 characters, must contain only alphanumeric characters or hyphens (-), and must not begin or end with a hyphen.

idempotencyToken :: Text

A unique, case-sensitive identifier that you provide to ensure the idempotency—that is, the uniqueness—of the request.

newCreateCustomRoutingAccelerator

Arguments

:: Text
$sel:name:CreateCustomRoutingAccelerator'

-> Text
$sel:idempotencyToken:CreateCustomRoutingAccelerator'

-> CreateCustomRoutingAccelerator

Create a value of CreateCustomRoutingAccelerator with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:enabled:CreateCustomRoutingAccelerator', createCustomRoutingAccelerator_enabled - Indicates whether an accelerator is enabled. The value is true or false. The default value is true.

If the value is set to true, an accelerator cannot be deleted. If set to false, the accelerator can be deleted.

$sel:ipAddressType:CreateCustomRoutingAccelerator', createCustomRoutingAccelerator_ipAddressType - The value for the address type must be IPv4.

$sel:tags:CreateCustomRoutingAccelerator', createCustomRoutingAccelerator_tags - Create tags for an accelerator.

For more information, see Tagging in AWS Global Accelerator in the AWS Global Accelerator Developer Guide.

$sel:ipAddresses:CreateCustomRoutingAccelerator', createCustomRoutingAccelerator_ipAddresses - Optionally, if you've added your own IP address pool to Global Accelerator (BYOIP), you can choose IP addresses from your own pool to use for the accelerator's static IP addresses when you create an accelerator. You can specify one or two addresses, separated by a space. Do not include the /32 suffix.

Only one IP address from each of your IP address ranges can be used for each accelerator. If you specify only one IP address from your IP address range, Global Accelerator assigns a second static IP address for the accelerator from the AWS IP address pool.

Note that you can't update IP addresses for an existing accelerator. To change them, you must create a new accelerator with the new addresses.

For more information, see Bring your own IP addresses (BYOIP) in the AWS Global Accelerator Developer Guide.

$sel:name:CreateCustomRoutingAccelerator', createCustomRoutingAccelerator_name - The name of a custom routing accelerator. The name can have a maximum of 64 characters, must contain only alphanumeric characters or hyphens (-), and must not begin or end with a hyphen.

$sel:idempotencyToken:CreateCustomRoutingAccelerator', createCustomRoutingAccelerator_idempotencyToken - A unique, case-sensitive identifier that you provide to ensure the idempotency—that is, the uniqueness—of the request.

Request Lenses

createCustomRoutingAccelerator_enabled :: Lens' CreateCustomRoutingAccelerator (Maybe Bool)

Indicates whether an accelerator is enabled. The value is true or false. The default value is true.

If the value is set to true, an accelerator cannot be deleted. If set to false, the accelerator can be deleted.

createCustomRoutingAccelerator_ipAddressType :: Lens' CreateCustomRoutingAccelerator (Maybe IpAddressType)

The value for the address type must be IPv4.

createCustomRoutingAccelerator_tags :: Lens' CreateCustomRoutingAccelerator (Maybe [Tag])

Create tags for an accelerator.

For more information, see Tagging in AWS Global Accelerator in the AWS Global Accelerator Developer Guide.

createCustomRoutingAccelerator_ipAddresses :: Lens' CreateCustomRoutingAccelerator (Maybe [Text])

Optionally, if you've added your own IP address pool to Global Accelerator (BYOIP), you can choose IP addresses from your own pool to use for the accelerator's static IP addresses when you create an accelerator. You can specify one or two addresses, separated by a space. Do not include the /32 suffix.

Only one IP address from each of your IP address ranges can be used for each accelerator. If you specify only one IP address from your IP address range, Global Accelerator assigns a second static IP address for the accelerator from the AWS IP address pool.

Note that you can't update IP addresses for an existing accelerator. To change them, you must create a new accelerator with the new addresses.

For more information, see Bring your own IP addresses (BYOIP) in the AWS Global Accelerator Developer Guide.

createCustomRoutingAccelerator_name :: Lens' CreateCustomRoutingAccelerator Text

The name of a custom routing accelerator. The name can have a maximum of 64 characters, must contain only alphanumeric characters or hyphens (-), and must not begin or end with a hyphen.

createCustomRoutingAccelerator_idempotencyToken :: Lens' CreateCustomRoutingAccelerator Text

A unique, case-sensitive identifier that you provide to ensure the idempotency—that is, the uniqueness—of the request.

Destructuring the Response

data CreateCustomRoutingAcceleratorResponse

See: newCreateCustomRoutingAcceleratorResponse smart constructor.

Constructors

CreateCustomRoutingAcceleratorResponse'

Fields

accelerator :: Maybe CustomRoutingAccelerator

The accelerator that is created.

httpStatus :: Int

The response's http status code.

newCreateCustomRoutingAcceleratorResponse

Arguments

:: Int
$sel:httpStatus:CreateCustomRoutingAcceleratorResponse'

-> CreateCustomRoutingAcceleratorResponse

Create a value of CreateCustomRoutingAcceleratorResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:accelerator:CreateCustomRoutingAcceleratorResponse', createCustomRoutingAcceleratorResponse_accelerator - The accelerator that is created.

$sel:httpStatus:CreateCustomRoutingAcceleratorResponse', createCustomRoutingAcceleratorResponse_httpStatus - The response's http status code.

Response Lenses

createCustomRoutingAcceleratorResponse_accelerator :: Lens' CreateCustomRoutingAcceleratorResponse (Maybe CustomRoutingAccelerator)

The accelerator that is created.

createCustomRoutingAcceleratorResponse_httpStatus :: Lens' CreateCustomRoutingAcceleratorResponse Int

The response's http status code.