Amazonka.GlobalAccelerator.CreateEndpointGroup

Description

Create an endpoint group for the specified listener. An endpoint group is a collection of endpoints in one AWS Region. A resource must be valid and active when you add it as an endpoint.

Creating a Request

data CreateEndpointGroup

See: newCreateEndpointGroup smart constructor.

Constructors

CreateEndpointGroup'

Fields

thresholdCount :: Maybe Natural

The number of consecutive health checks required to set the state of a healthy endpoint to unhealthy, or to set an unhealthy endpoint to healthy. The default value is 3.

healthCheckPath :: Maybe Text

If the protocol is HTTP/S, then this specifies the path that is the destination for health check targets. The default value is slash (/).

healthCheckIntervalSeconds :: Maybe Natural

The time—10 seconds or 30 seconds—between each health check for an endpoint. The default value is 30.

healthCheckProtocol :: Maybe HealthCheckProtocol

The protocol that AWS Global Accelerator uses to check the health of endpoints that are part of this endpoint group. The default value is TCP.

trafficDialPercentage :: Maybe Double

The percentage of traffic to send to an AWS Region. Additional traffic is distributed to other endpoint groups for this listener.

Use this action to increase (dial up) or decrease (dial down) traffic to a specific Region. The percentage is applied to the traffic that would otherwise have been routed to the Region based on optimal routing.

The default value is 100.

endpointConfigurations :: Maybe [EndpointConfiguration]

The list of endpoint objects.

healthCheckPort :: Maybe Natural

The port that AWS Global Accelerator uses to check the health of endpoints that are part of this endpoint group. The default port is the listener port that this endpoint group is associated with. If listener port is a list of ports, Global Accelerator uses the first port in the list.

portOverrides :: Maybe [PortOverride]

Override specific listener ports used to route traffic to endpoints that are part of this endpoint group. For example, you can create a port override in which the listener receives user traffic on ports 80 and 443, but your accelerator routes that traffic to ports 1080 and 1443, respectively, on the endpoints.

For more information, see Port overrides in the AWS Global Accelerator Developer Guide.

listenerArn :: Text

The Amazon Resource Name (ARN) of the listener.

endpointGroupRegion :: Text

The AWS Region where the endpoint group is located. A listener can have only one endpoint group in a specific Region.

idempotencyToken :: Text

A unique, case-sensitive identifier that you provide to ensure the idempotency—that is, the uniqueness—of the request.

newCreateEndpointGroup

Arguments

:: Text
$sel:listenerArn:CreateEndpointGroup'

-> Text
$sel:endpointGroupRegion:CreateEndpointGroup'

-> Text
$sel:idempotencyToken:CreateEndpointGroup'

-> CreateEndpointGroup

Create a value of CreateEndpointGroup with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:thresholdCount:CreateEndpointGroup', createEndpointGroup_thresholdCount - The number of consecutive health checks required to set the state of a healthy endpoint to unhealthy, or to set an unhealthy endpoint to healthy. The default value is 3.

$sel:healthCheckPath:CreateEndpointGroup', createEndpointGroup_healthCheckPath - If the protocol is HTTP/S, then this specifies the path that is the destination for health check targets. The default value is slash (/).

$sel:healthCheckIntervalSeconds:CreateEndpointGroup', createEndpointGroup_healthCheckIntervalSeconds - The time—10 seconds or 30 seconds—between each health check for an endpoint. The default value is 30.

$sel:healthCheckProtocol:CreateEndpointGroup', createEndpointGroup_healthCheckProtocol - The protocol that AWS Global Accelerator uses to check the health of endpoints that are part of this endpoint group. The default value is TCP.

$sel:trafficDialPercentage:CreateEndpointGroup', createEndpointGroup_trafficDialPercentage - The percentage of traffic to send to an AWS Region. Additional traffic is distributed to other endpoint groups for this listener.

Use this action to increase (dial up) or decrease (dial down) traffic to a specific Region. The percentage is applied to the traffic that would otherwise have been routed to the Region based on optimal routing.

The default value is 100.

$sel:endpointConfigurations:CreateEndpointGroup', createEndpointGroup_endpointConfigurations - The list of endpoint objects.

$sel:healthCheckPort:CreateEndpointGroup', createEndpointGroup_healthCheckPort - The port that AWS Global Accelerator uses to check the health of endpoints that are part of this endpoint group. The default port is the listener port that this endpoint group is associated with. If listener port is a list of ports, Global Accelerator uses the first port in the list.

$sel:portOverrides:CreateEndpointGroup', createEndpointGroup_portOverrides - Override specific listener ports used to route traffic to endpoints that are part of this endpoint group. For example, you can create a port override in which the listener receives user traffic on ports 80 and 443, but your accelerator routes that traffic to ports 1080 and 1443, respectively, on the endpoints.

For more information, see Port overrides in the AWS Global Accelerator Developer Guide.

$sel:listenerArn:CreateEndpointGroup', createEndpointGroup_listenerArn - The Amazon Resource Name (ARN) of the listener.

$sel:endpointGroupRegion:CreateEndpointGroup', createEndpointGroup_endpointGroupRegion - The AWS Region where the endpoint group is located. A listener can have only one endpoint group in a specific Region.

$sel:idempotencyToken:CreateEndpointGroup', createEndpointGroup_idempotencyToken - A unique, case-sensitive identifier that you provide to ensure the idempotency—that is, the uniqueness—of the request.

Request Lenses

createEndpointGroup_thresholdCount :: Lens' CreateEndpointGroup (Maybe Natural)

The number of consecutive health checks required to set the state of a healthy endpoint to unhealthy, or to set an unhealthy endpoint to healthy. The default value is 3.

createEndpointGroup_healthCheckPath :: Lens' CreateEndpointGroup (Maybe Text)

If the protocol is HTTP/S, then this specifies the path that is the destination for health check targets. The default value is slash (/).

createEndpointGroup_healthCheckIntervalSeconds :: Lens' CreateEndpointGroup (Maybe Natural)

The time—10 seconds or 30 seconds—between each health check for an endpoint. The default value is 30.

createEndpointGroup_healthCheckProtocol :: Lens' CreateEndpointGroup (Maybe HealthCheckProtocol)

The protocol that AWS Global Accelerator uses to check the health of endpoints that are part of this endpoint group. The default value is TCP.

createEndpointGroup_trafficDialPercentage :: Lens' CreateEndpointGroup (Maybe Double)

The percentage of traffic to send to an AWS Region. Additional traffic is distributed to other endpoint groups for this listener.

Use this action to increase (dial up) or decrease (dial down) traffic to a specific Region. The percentage is applied to the traffic that would otherwise have been routed to the Region based on optimal routing.

The default value is 100.

createEndpointGroup_endpointConfigurations :: Lens' CreateEndpointGroup (Maybe [EndpointConfiguration])

The list of endpoint objects.

createEndpointGroup_healthCheckPort :: Lens' CreateEndpointGroup (Maybe Natural)

The port that AWS Global Accelerator uses to check the health of endpoints that are part of this endpoint group. The default port is the listener port that this endpoint group is associated with. If listener port is a list of ports, Global Accelerator uses the first port in the list.

createEndpointGroup_portOverrides :: Lens' CreateEndpointGroup (Maybe [PortOverride])

Override specific listener ports used to route traffic to endpoints that are part of this endpoint group. For example, you can create a port override in which the listener receives user traffic on ports 80 and 443, but your accelerator routes that traffic to ports 1080 and 1443, respectively, on the endpoints.

For more information, see Port overrides in the AWS Global Accelerator Developer Guide.

createEndpointGroup_listenerArn :: Lens' CreateEndpointGroup Text

The Amazon Resource Name (ARN) of the listener.

createEndpointGroup_endpointGroupRegion :: Lens' CreateEndpointGroup Text

The AWS Region where the endpoint group is located. A listener can have only one endpoint group in a specific Region.

createEndpointGroup_idempotencyToken :: Lens' CreateEndpointGroup Text

A unique, case-sensitive identifier that you provide to ensure the idempotency—that is, the uniqueness—of the request.

Destructuring the Response

data CreateEndpointGroupResponse

See: newCreateEndpointGroupResponse smart constructor.

Constructors

CreateEndpointGroupResponse'

Fields

endpointGroup :: Maybe EndpointGroup

The information about the endpoint group that was created.

httpStatus :: Int

The response's http status code.

newCreateEndpointGroupResponse

Arguments

:: Int
$sel:httpStatus:CreateEndpointGroupResponse'

-> CreateEndpointGroupResponse

Create a value of CreateEndpointGroupResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:endpointGroup:CreateEndpointGroupResponse', createEndpointGroupResponse_endpointGroup - The information about the endpoint group that was created.

$sel:httpStatus:CreateEndpointGroupResponse', createEndpointGroupResponse_httpStatus - The response's http status code.

Response Lenses

createEndpointGroupResponse_endpointGroup :: Lens' CreateEndpointGroupResponse (Maybe EndpointGroup)

The information about the endpoint group that was created.

createEndpointGroupResponse_httpStatus :: Lens' CreateEndpointGroupResponse Int

The response's http status code.