Amazonka.GlobalAccelerator.DenyCustomRoutingTraffic

Description

Specify the Amazon EC2 instance (destination) IP addresses and ports for a VPC subnet endpoint that cannot receive traffic for a custom routing accelerator. You can deny traffic to all destinations in the VPC endpoint, or deny traffic to a specified list of destination IP addresses and ports. Note that you cannot specify IP addresses or ports outside of the range that you configured for the endpoint group.

After you make changes, you can verify that the updates are complete by checking the status of your accelerator: the status changes from IN_PROGRESS to DEPLOYED.

Creating a Request

data DenyCustomRoutingTraffic

See: newDenyCustomRoutingTraffic smart constructor.

Constructors

DenyCustomRoutingTraffic'

Fields

destinationAddresses :: Maybe [Text]

A list of specific Amazon EC2 instance IP addresses (destination addresses) in a subnet that you want to prevent from receiving traffic. The IP addresses must be a subset of the IP addresses allowed for the VPC subnet associated with the endpoint group.

destinationPorts :: Maybe [Natural]

A list of specific Amazon EC2 instance ports (destination ports) in a subnet endpoint that you want to prevent from receiving traffic.

denyAllTrafficToEndpoint :: Maybe Bool

Indicates whether all destination IP addresses and ports for a specified VPC subnet endpoint cannot receive traffic from a custom routing accelerator. The value is TRUE or FALSE.

When set to TRUE, no destinations in the custom routing VPC subnet can receive traffic. Note that you cannot specify destination IP addresses and ports when the value is set to TRUE.

When set to FALSE (or not specified), you must specify a list of destination IP addresses that cannot receive traffic. A list of ports is optional. If you don't specify a list of ports, the ports that can accept traffic is the same as the ports configured for the endpoint group.

The default value is FALSE.

endpointGroupArn :: Text

The Amazon Resource Name (ARN) of the endpoint group.

endpointId :: Text

An ID for the endpoint. For custom routing accelerators, this is the virtual private cloud (VPC) subnet ID.

newDenyCustomRoutingTraffic

Arguments

:: Text
$sel:endpointGroupArn:DenyCustomRoutingTraffic'

-> Text
$sel:endpointId:DenyCustomRoutingTraffic'

-> DenyCustomRoutingTraffic

Create a value of DenyCustomRoutingTraffic with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:destinationAddresses:DenyCustomRoutingTraffic', denyCustomRoutingTraffic_destinationAddresses - A list of specific Amazon EC2 instance IP addresses (destination addresses) in a subnet that you want to prevent from receiving traffic. The IP addresses must be a subset of the IP addresses allowed for the VPC subnet associated with the endpoint group.

$sel:destinationPorts:DenyCustomRoutingTraffic', denyCustomRoutingTraffic_destinationPorts - A list of specific Amazon EC2 instance ports (destination ports) in a subnet endpoint that you want to prevent from receiving traffic.

$sel:denyAllTrafficToEndpoint:DenyCustomRoutingTraffic', denyCustomRoutingTraffic_denyAllTrafficToEndpoint - Indicates whether all destination IP addresses and ports for a specified VPC subnet endpoint cannot receive traffic from a custom routing accelerator. The value is TRUE or FALSE.

When set to TRUE, no destinations in the custom routing VPC subnet can receive traffic. Note that you cannot specify destination IP addresses and ports when the value is set to TRUE.

When set to FALSE (or not specified), you must specify a list of destination IP addresses that cannot receive traffic. A list of ports is optional. If you don't specify a list of ports, the ports that can accept traffic is the same as the ports configured for the endpoint group.

The default value is FALSE.

$sel:endpointGroupArn:DenyCustomRoutingTraffic', denyCustomRoutingTraffic_endpointGroupArn - The Amazon Resource Name (ARN) of the endpoint group.

$sel:endpointId:DenyCustomRoutingTraffic', denyCustomRoutingTraffic_endpointId - An ID for the endpoint. For custom routing accelerators, this is the virtual private cloud (VPC) subnet ID.

Request Lenses

denyCustomRoutingTraffic_destinationAddresses :: Lens' DenyCustomRoutingTraffic (Maybe [Text])

A list of specific Amazon EC2 instance IP addresses (destination addresses) in a subnet that you want to prevent from receiving traffic. The IP addresses must be a subset of the IP addresses allowed for the VPC subnet associated with the endpoint group.

denyCustomRoutingTraffic_destinationPorts :: Lens' DenyCustomRoutingTraffic (Maybe [Natural])

A list of specific Amazon EC2 instance ports (destination ports) in a subnet endpoint that you want to prevent from receiving traffic.

denyCustomRoutingTraffic_denyAllTrafficToEndpoint :: Lens' DenyCustomRoutingTraffic (Maybe Bool)

Indicates whether all destination IP addresses and ports for a specified VPC subnet endpoint cannot receive traffic from a custom routing accelerator. The value is TRUE or FALSE.

When set to TRUE, no destinations in the custom routing VPC subnet can receive traffic. Note that you cannot specify destination IP addresses and ports when the value is set to TRUE.

When set to FALSE (or not specified), you must specify a list of destination IP addresses that cannot receive traffic. A list of ports is optional. If you don't specify a list of ports, the ports that can accept traffic is the same as the ports configured for the endpoint group.

The default value is FALSE.

denyCustomRoutingTraffic_endpointGroupArn :: Lens' DenyCustomRoutingTraffic Text

The Amazon Resource Name (ARN) of the endpoint group.

denyCustomRoutingTraffic_endpointId :: Lens' DenyCustomRoutingTraffic Text

An ID for the endpoint. For custom routing accelerators, this is the virtual private cloud (VPC) subnet ID.

Destructuring the Response

data DenyCustomRoutingTrafficResponse

See: newDenyCustomRoutingTrafficResponse smart constructor.

Constructors

DenyCustomRoutingTrafficResponse'

newDenyCustomRoutingTrafficResponse :: DenyCustomRoutingTrafficResponse

Create a value of DenyCustomRoutingTrafficResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.