Amazonka.GlobalAccelerator.DescribeCustomRoutingAccelerator

Description

Describe a custom routing accelerator.

Creating a Request

data DescribeCustomRoutingAccelerator

See: newDescribeCustomRoutingAccelerator smart constructor.

Constructors

DescribeCustomRoutingAccelerator'

Fields

acceleratorArn :: Text

The Amazon Resource Name (ARN) of the accelerator to describe.

newDescribeCustomRoutingAccelerator

Arguments

:: Text
$sel:acceleratorArn:DescribeCustomRoutingAccelerator'

-> DescribeCustomRoutingAccelerator

Create a value of DescribeCustomRoutingAccelerator with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:acceleratorArn:DescribeCustomRoutingAccelerator', describeCustomRoutingAccelerator_acceleratorArn - The Amazon Resource Name (ARN) of the accelerator to describe.

Request Lenses

describeCustomRoutingAccelerator_acceleratorArn :: Lens' DescribeCustomRoutingAccelerator Text

The Amazon Resource Name (ARN) of the accelerator to describe.

Destructuring the Response

data DescribeCustomRoutingAcceleratorResponse

See: newDescribeCustomRoutingAcceleratorResponse smart constructor.

Constructors

DescribeCustomRoutingAcceleratorResponse'

Fields

accelerator :: Maybe CustomRoutingAccelerator

The description of the custom routing accelerator.

httpStatus :: Int

The response's http status code.

newDescribeCustomRoutingAcceleratorResponse

Arguments

:: Int
$sel:httpStatus:DescribeCustomRoutingAcceleratorResponse'

-> DescribeCustomRoutingAcceleratorResponse

Create a value of DescribeCustomRoutingAcceleratorResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:accelerator:DescribeCustomRoutingAcceleratorResponse', describeCustomRoutingAcceleratorResponse_accelerator - The description of the custom routing accelerator.

$sel:httpStatus:DescribeCustomRoutingAcceleratorResponse', describeCustomRoutingAcceleratorResponse_httpStatus - The response's http status code.

Response Lenses

describeCustomRoutingAcceleratorResponse_accelerator :: Lens' DescribeCustomRoutingAcceleratorResponse (Maybe CustomRoutingAccelerator)

The description of the custom routing accelerator.

describeCustomRoutingAcceleratorResponse_httpStatus :: Lens' DescribeCustomRoutingAcceleratorResponse Int

The response's http status code.