Amazonka.GlobalAccelerator.ListCustomRoutingPortMappings

Description

Provides a complete mapping from the public accelerator IP address and port to destination EC2 instance IP addresses and ports in the virtual public cloud (VPC) subnet endpoint for a custom routing accelerator. For each subnet endpoint that you add, Global Accelerator creates a new static port mapping for the accelerator. The port mappings don't change after Global Accelerator generates them, so you can retrieve and cache the full mapping on your servers.

If you remove a subnet from your accelerator, Global Accelerator removes (reclaims) the port mappings. If you add a subnet to your accelerator, Global Accelerator creates new port mappings (the existing ones don't change). If you add or remove EC2 instances in your subnet, the port mappings don't change, because the mappings are created when you add the subnet to Global Accelerator.

The mappings also include a flag for each destination denoting which destination IP addresses and ports are allowed or denied traffic.

This operation returns paginated results.

Creating a Request

data ListCustomRoutingPortMappings

See: newListCustomRoutingPortMappings smart constructor.

Constructors

ListCustomRoutingPortMappings'

Fields

endpointGroupArn :: Maybe Text

The Amazon Resource Name (ARN) of the endpoint group to list the custom routing port mappings for.

nextToken :: Maybe Text

The token for the next set of results. You receive this token from a previous call.

maxResults :: Maybe Natural

The number of destination port mappings that you want to return with this call. The default value is 10.

acceleratorArn :: Text

The Amazon Resource Name (ARN) of the accelerator to list the custom routing port mappings for.

newListCustomRoutingPortMappings

Arguments

:: Text
$sel:acceleratorArn:ListCustomRoutingPortMappings'

-> ListCustomRoutingPortMappings

Create a value of ListCustomRoutingPortMappings with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:endpointGroupArn:ListCustomRoutingPortMappings', listCustomRoutingPortMappings_endpointGroupArn - The Amazon Resource Name (ARN) of the endpoint group to list the custom routing port mappings for.

$sel:nextToken:ListCustomRoutingPortMappings', listCustomRoutingPortMappings_nextToken - The token for the next set of results. You receive this token from a previous call.

$sel:maxResults:ListCustomRoutingPortMappings', listCustomRoutingPortMappings_maxResults - The number of destination port mappings that you want to return with this call. The default value is 10.

$sel:acceleratorArn:ListCustomRoutingPortMappings', listCustomRoutingPortMappings_acceleratorArn - The Amazon Resource Name (ARN) of the accelerator to list the custom routing port mappings for.

Request Lenses

listCustomRoutingPortMappings_endpointGroupArn :: Lens' ListCustomRoutingPortMappings (Maybe Text)

The Amazon Resource Name (ARN) of the endpoint group to list the custom routing port mappings for.

listCustomRoutingPortMappings_nextToken :: Lens' ListCustomRoutingPortMappings (Maybe Text)

The token for the next set of results. You receive this token from a previous call.

listCustomRoutingPortMappings_maxResults :: Lens' ListCustomRoutingPortMappings (Maybe Natural)

The number of destination port mappings that you want to return with this call. The default value is 10.

listCustomRoutingPortMappings_acceleratorArn :: Lens' ListCustomRoutingPortMappings Text

The Amazon Resource Name (ARN) of the accelerator to list the custom routing port mappings for.

Destructuring the Response

data ListCustomRoutingPortMappingsResponse

See: newListCustomRoutingPortMappingsResponse smart constructor.

Constructors

ListCustomRoutingPortMappingsResponse'

Fields

portMappings :: Maybe [PortMapping]

The port mappings for a custom routing accelerator.

nextToken :: Maybe Text

The token for the next set of results. You receive this token from a previous call.

httpStatus :: Int

The response's http status code.

newListCustomRoutingPortMappingsResponse

Arguments

:: Int
$sel:httpStatus:ListCustomRoutingPortMappingsResponse'

-> ListCustomRoutingPortMappingsResponse

Create a value of ListCustomRoutingPortMappingsResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:portMappings:ListCustomRoutingPortMappingsResponse', listCustomRoutingPortMappingsResponse_portMappings - The port mappings for a custom routing accelerator.

$sel:nextToken:ListCustomRoutingPortMappings', listCustomRoutingPortMappingsResponse_nextToken - The token for the next set of results. You receive this token from a previous call.

$sel:httpStatus:ListCustomRoutingPortMappingsResponse', listCustomRoutingPortMappingsResponse_httpStatus - The response's http status code.

Response Lenses

listCustomRoutingPortMappingsResponse_portMappings :: Lens' ListCustomRoutingPortMappingsResponse (Maybe [PortMapping])

The port mappings for a custom routing accelerator.

listCustomRoutingPortMappingsResponse_nextToken :: Lens' ListCustomRoutingPortMappingsResponse (Maybe Text)

The token for the next set of results. You receive this token from a previous call.

listCustomRoutingPortMappingsResponse_httpStatus :: Lens' ListCustomRoutingPortMappingsResponse Int

The response's http status code.