Amazonka.GlobalAccelerator.ListCustomRoutingPortMappingsByDestination

Description

List the port mappings for a specific EC2 instance (destination) in a VPC subnet endpoint. The response is the mappings for one destination IP address. This is useful when your subnet endpoint has mappings that span multiple custom routing accelerators in your account, or for scenarios where you only want to list the port mappings for a specific destination instance.

This operation returns paginated results.

Creating a Request

data ListCustomRoutingPortMappingsByDestination

See: newListCustomRoutingPortMappingsByDestination smart constructor.

Constructors

ListCustomRoutingPortMappingsByDestination'

Fields

nextToken :: Maybe Text

The token for the next set of results. You receive this token from a previous call.

maxResults :: Maybe Natural

The number of destination port mappings that you want to return with this call. The default value is 10.

endpointId :: Text

The ID for the virtual private cloud (VPC) subnet.

destinationAddress :: Text

The endpoint IP address in a virtual private cloud (VPC) subnet for which you want to receive back port mappings.

newListCustomRoutingPortMappingsByDestination

Arguments

:: Text
$sel:endpointId:ListCustomRoutingPortMappingsByDestination'

-> Text
$sel:destinationAddress:ListCustomRoutingPortMappingsByDestination'

-> ListCustomRoutingPortMappingsByDestination

Create a value of ListCustomRoutingPortMappingsByDestination with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:nextToken:ListCustomRoutingPortMappingsByDestination', listCustomRoutingPortMappingsByDestination_nextToken - The token for the next set of results. You receive this token from a previous call.

$sel:maxResults:ListCustomRoutingPortMappingsByDestination', listCustomRoutingPortMappingsByDestination_maxResults - The number of destination port mappings that you want to return with this call. The default value is 10.

$sel:endpointId:ListCustomRoutingPortMappingsByDestination', listCustomRoutingPortMappingsByDestination_endpointId - The ID for the virtual private cloud (VPC) subnet.

$sel:destinationAddress:ListCustomRoutingPortMappingsByDestination', listCustomRoutingPortMappingsByDestination_destinationAddress - The endpoint IP address in a virtual private cloud (VPC) subnet for which you want to receive back port mappings.

Request Lenses

listCustomRoutingPortMappingsByDestination_nextToken :: Lens' ListCustomRoutingPortMappingsByDestination (Maybe Text)

The token for the next set of results. You receive this token from a previous call.

listCustomRoutingPortMappingsByDestination_maxResults :: Lens' ListCustomRoutingPortMappingsByDestination (Maybe Natural)

The number of destination port mappings that you want to return with this call. The default value is 10.

listCustomRoutingPortMappingsByDestination_endpointId :: Lens' ListCustomRoutingPortMappingsByDestination Text

The ID for the virtual private cloud (VPC) subnet.

listCustomRoutingPortMappingsByDestination_destinationAddress :: Lens' ListCustomRoutingPortMappingsByDestination Text

The endpoint IP address in a virtual private cloud (VPC) subnet for which you want to receive back port mappings.

Destructuring the Response

data ListCustomRoutingPortMappingsByDestinationResponse

See: newListCustomRoutingPortMappingsByDestinationResponse smart constructor.

Constructors

ListCustomRoutingPortMappingsByDestinationResponse'

Fields

nextToken :: Maybe Text

The token for the next set of results. You receive this token from a previous call.

destinationPortMappings :: Maybe [DestinationPortMapping]

The port mappings for the endpoint IP address that you specified in the request.

httpStatus :: Int

The response's http status code.

newListCustomRoutingPortMappingsByDestinationResponse

Arguments

:: Int
$sel:httpStatus:ListCustomRoutingPortMappingsByDestinationResponse'

-> ListCustomRoutingPortMappingsByDestinationResponse

Create a value of ListCustomRoutingPortMappingsByDestinationResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:nextToken:ListCustomRoutingPortMappingsByDestination', listCustomRoutingPortMappingsByDestinationResponse_nextToken - The token for the next set of results. You receive this token from a previous call.

$sel:destinationPortMappings:ListCustomRoutingPortMappingsByDestinationResponse', listCustomRoutingPortMappingsByDestinationResponse_destinationPortMappings - The port mappings for the endpoint IP address that you specified in the request.

$sel:httpStatus:ListCustomRoutingPortMappingsByDestinationResponse', listCustomRoutingPortMappingsByDestinationResponse_httpStatus - The response's http status code.

Response Lenses

listCustomRoutingPortMappingsByDestinationResponse_nextToken :: Lens' ListCustomRoutingPortMappingsByDestinationResponse (Maybe Text)

The token for the next set of results. You receive this token from a previous call.

listCustomRoutingPortMappingsByDestinationResponse_destinationPortMappings :: Lens' ListCustomRoutingPortMappingsByDestinationResponse (Maybe [DestinationPortMapping])

The port mappings for the endpoint IP address that you specified in the request.

listCustomRoutingPortMappingsByDestinationResponse_httpStatus :: Lens' ListCustomRoutingPortMappingsByDestinationResponse Int

The response's http status code.