Amazonka.GlobalAccelerator.Types.CustomRoutingAccelerator

Description

Documentation

data CustomRoutingAccelerator

Attributes of a custom routing accelerator.

See: newCustomRoutingAccelerator smart constructor.

Constructors

CustomRoutingAccelerator'

Fields

status :: Maybe CustomRoutingAcceleratorStatus

Describes the deployment status of the accelerator.

acceleratorArn :: Maybe Text

The Amazon Resource Name (ARN) of the custom routing accelerator.

enabled :: Maybe Bool

Indicates whether the accelerator is enabled. The value is true or false. The default value is true.

If the value is set to true, the accelerator cannot be deleted. If set to false, accelerator can be deleted.

createdTime :: Maybe POSIX

The date and time that the accelerator was created.

lastModifiedTime :: Maybe POSIX

The date and time that the accelerator was last modified.

ipAddressType :: Maybe IpAddressType

The value for the address type must be IPv4.

name :: Maybe Text

The name of the accelerator. The name must contain only alphanumeric characters or hyphens (-), and must not begin or end with a hyphen.

ipSets :: Maybe [IpSet]

The static IP addresses that Global Accelerator associates with the accelerator.

dnsName :: Maybe Text

The Domain Name System (DNS) name that Global Accelerator creates that points to your accelerator's static IP addresses.

The naming convention for the DNS name is the following: A lowercase letter a, followed by a 16-bit random hex string, followed by .awsglobalaccelerator.com. For example: a1234567890abcdef.awsglobalaccelerator.com.

For more information about the default DNS name, see Support for DNS Addressing in Global Accelerator in the AWS Global Accelerator Developer Guide.

newCustomRoutingAccelerator :: CustomRoutingAccelerator

Create a value of CustomRoutingAccelerator with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:status:CustomRoutingAccelerator', customRoutingAccelerator_status - Describes the deployment status of the accelerator.

$sel:acceleratorArn:CustomRoutingAccelerator', customRoutingAccelerator_acceleratorArn - The Amazon Resource Name (ARN) of the custom routing accelerator.

$sel:enabled:CustomRoutingAccelerator', customRoutingAccelerator_enabled - Indicates whether the accelerator is enabled. The value is true or false. The default value is true.

If the value is set to true, the accelerator cannot be deleted. If set to false, accelerator can be deleted.

$sel:createdTime:CustomRoutingAccelerator', customRoutingAccelerator_createdTime - The date and time that the accelerator was created.

$sel:lastModifiedTime:CustomRoutingAccelerator', customRoutingAccelerator_lastModifiedTime - The date and time that the accelerator was last modified.

$sel:ipAddressType:CustomRoutingAccelerator', customRoutingAccelerator_ipAddressType - The value for the address type must be IPv4.

$sel:name:CustomRoutingAccelerator', customRoutingAccelerator_name - The name of the accelerator. The name must contain only alphanumeric characters or hyphens (-), and must not begin or end with a hyphen.

$sel:ipSets:CustomRoutingAccelerator', customRoutingAccelerator_ipSets - The static IP addresses that Global Accelerator associates with the accelerator.

$sel:dnsName:CustomRoutingAccelerator', customRoutingAccelerator_dnsName - The Domain Name System (DNS) name that Global Accelerator creates that points to your accelerator's static IP addresses.

The naming convention for the DNS name is the following: A lowercase letter a, followed by a 16-bit random hex string, followed by .awsglobalaccelerator.com. For example: a1234567890abcdef.awsglobalaccelerator.com.

For more information about the default DNS name, see Support for DNS Addressing in Global Accelerator in the AWS Global Accelerator Developer Guide.

customRoutingAccelerator_status :: Lens' CustomRoutingAccelerator (Maybe CustomRoutingAcceleratorStatus)

Describes the deployment status of the accelerator.

customRoutingAccelerator_acceleratorArn :: Lens' CustomRoutingAccelerator (Maybe Text)

The Amazon Resource Name (ARN) of the custom routing accelerator.

customRoutingAccelerator_enabled :: Lens' CustomRoutingAccelerator (Maybe Bool)

Indicates whether the accelerator is enabled. The value is true or false. The default value is true.

If the value is set to true, the accelerator cannot be deleted. If set to false, accelerator can be deleted.

customRoutingAccelerator_createdTime :: Lens' CustomRoutingAccelerator (Maybe UTCTime)

The date and time that the accelerator was created.

customRoutingAccelerator_lastModifiedTime :: Lens' CustomRoutingAccelerator (Maybe UTCTime)

The date and time that the accelerator was last modified.

customRoutingAccelerator_ipAddressType :: Lens' CustomRoutingAccelerator (Maybe IpAddressType)

The value for the address type must be IPv4.

customRoutingAccelerator_name :: Lens' CustomRoutingAccelerator (Maybe Text)

The name of the accelerator. The name must contain only alphanumeric characters or hyphens (-), and must not begin or end with a hyphen.

customRoutingAccelerator_ipSets :: Lens' CustomRoutingAccelerator (Maybe [IpSet])

The static IP addresses that Global Accelerator associates with the accelerator.

customRoutingAccelerator_dnsName :: Lens' CustomRoutingAccelerator (Maybe Text)

The Domain Name System (DNS) name that Global Accelerator creates that points to your accelerator's static IP addresses.

The naming convention for the DNS name is the following: A lowercase letter a, followed by a 16-bit random hex string, followed by .awsglobalaccelerator.com. For example: a1234567890abcdef.awsglobalaccelerator.com.

For more information about the default DNS name, see Support for DNS Addressing in Global Accelerator in the AWS Global Accelerator Developer Guide.