Amazonka.GlobalAccelerator.Types.DestinationPortMapping

Description

Documentation

data DestinationPortMapping

The port mappings for a specified endpoint IP address (destination).

See: newDestinationPortMapping smart constructor.

Constructors

DestinationPortMapping'

Fields

destinationSocketAddress :: Maybe SocketAddress

The endpoint IP address/port combination for traffic received on the accelerator socket address.

acceleratorArn :: Maybe Text

The Amazon Resource Name (ARN) of the custom routing accelerator that you have port mappings for.

endpointGroupArn :: Maybe Text

The Amazon Resource Name (ARN) of the endpoint group.

endpointGroupRegion :: Maybe Text

The AWS Region for the endpoint group.

ipAddressType :: Maybe IpAddressType

The IP address type, which must be IPv4.

acceleratorSocketAddresses :: Maybe [SocketAddress]

The IP address/port combinations (sockets) that map to a given destination socket address.

endpointId :: Maybe Text

The ID for the virtual private cloud (VPC) subnet.

destinationTrafficState :: Maybe CustomRoutingDestinationTrafficState

Indicates whether or not a port mapping destination can receive traffic. The value is either ALLOW, if traffic is allowed to the destination, or DENY, if traffic is not allowed to the destination.

newDestinationPortMapping :: DestinationPortMapping

Create a value of DestinationPortMapping with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:destinationSocketAddress:DestinationPortMapping', destinationPortMapping_destinationSocketAddress - The endpoint IP address/port combination for traffic received on the accelerator socket address.

$sel:acceleratorArn:DestinationPortMapping', destinationPortMapping_acceleratorArn - The Amazon Resource Name (ARN) of the custom routing accelerator that you have port mappings for.

$sel:endpointGroupArn:DestinationPortMapping', destinationPortMapping_endpointGroupArn - The Amazon Resource Name (ARN) of the endpoint group.

$sel:endpointGroupRegion:DestinationPortMapping', destinationPortMapping_endpointGroupRegion - The AWS Region for the endpoint group.

$sel:ipAddressType:DestinationPortMapping', destinationPortMapping_ipAddressType - The IP address type, which must be IPv4.

$sel:acceleratorSocketAddresses:DestinationPortMapping', destinationPortMapping_acceleratorSocketAddresses - The IP address/port combinations (sockets) that map to a given destination socket address.

$sel:endpointId:DestinationPortMapping', destinationPortMapping_endpointId - The ID for the virtual private cloud (VPC) subnet.

$sel:destinationTrafficState:DestinationPortMapping', destinationPortMapping_destinationTrafficState - Indicates whether or not a port mapping destination can receive traffic. The value is either ALLOW, if traffic is allowed to the destination, or DENY, if traffic is not allowed to the destination.

destinationPortMapping_destinationSocketAddress :: Lens' DestinationPortMapping (Maybe SocketAddress)

The endpoint IP address/port combination for traffic received on the accelerator socket address.

destinationPortMapping_acceleratorArn :: Lens' DestinationPortMapping (Maybe Text)

The Amazon Resource Name (ARN) of the custom routing accelerator that you have port mappings for.

destinationPortMapping_endpointGroupArn :: Lens' DestinationPortMapping (Maybe Text)

The Amazon Resource Name (ARN) of the endpoint group.

destinationPortMapping_endpointGroupRegion :: Lens' DestinationPortMapping (Maybe Text)

The AWS Region for the endpoint group.

destinationPortMapping_ipAddressType :: Lens' DestinationPortMapping (Maybe IpAddressType)

The IP address type, which must be IPv4.

destinationPortMapping_acceleratorSocketAddresses :: Lens' DestinationPortMapping (Maybe [SocketAddress])

The IP address/port combinations (sockets) that map to a given destination socket address.

destinationPortMapping_endpointId :: Lens' DestinationPortMapping (Maybe Text)

The ID for the virtual private cloud (VPC) subnet.

destinationPortMapping_destinationTrafficState :: Lens' DestinationPortMapping (Maybe CustomRoutingDestinationTrafficState)

Indicates whether or not a port mapping destination can receive traffic. The value is either ALLOW, if traffic is allowed to the destination, or DENY, if traffic is not allowed to the destination.