Amazonka.GlobalAccelerator.Types.EndpointGroup

Description

Documentation

data EndpointGroup

A complex type for the endpoint group. An AWS Region can have only one endpoint group for a specific listener.

See: newEndpointGroup smart constructor.

Constructors

EndpointGroup'

Fields

thresholdCount :: Maybe Natural

The number of consecutive health checks required to set the state of a healthy endpoint to unhealthy, or to set an unhealthy endpoint to healthy. The default value is 3.

healthCheckPath :: Maybe Text

If the protocol is HTTP/S, then this value provides the ping path that Global Accelerator uses for the destination on the endpoints for health checks. The default is slash (/).

healthCheckIntervalSeconds :: Maybe Natural

The time—10 seconds or 30 seconds—between health checks for each endpoint. The default value is 30.

endpointGroupArn :: Maybe Text

The Amazon Resource Name (ARN) of the endpoint group.

healthCheckProtocol :: Maybe HealthCheckProtocol

The protocol that Global Accelerator uses to perform health checks on endpoints that are part of this endpoint group. The default value is TCP.

endpointGroupRegion :: Maybe Text

The AWS Region where the endpoint group is located.

trafficDialPercentage :: Maybe Double

The percentage of traffic to send to an AWS Region. Additional traffic is distributed to other endpoint groups for this listener.

Use this action to increase (dial up) or decrease (dial down) traffic to a specific Region. The percentage is applied to the traffic that would otherwise have been routed to the Region based on optimal routing.

The default value is 100.

healthCheckPort :: Maybe Natural

The port that Global Accelerator uses to perform health checks on endpoints that are part of this endpoint group.

The default port is the port for the listener that this endpoint group is associated with. If the listener port is a list, Global Accelerator uses the first specified port in the list of ports.

portOverrides :: Maybe [PortOverride]

Allows you to override the destination ports used to route traffic to an endpoint. Using a port override lets you to map a list of external destination ports (that your users send traffic to) to a list of internal destination ports that you want an application endpoint to receive traffic on.

endpointDescriptions :: Maybe [EndpointDescription]

The list of endpoint objects.

newEndpointGroup :: EndpointGroup

Create a value of EndpointGroup with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:thresholdCount:EndpointGroup', endpointGroup_thresholdCount - The number of consecutive health checks required to set the state of a healthy endpoint to unhealthy, or to set an unhealthy endpoint to healthy. The default value is 3.

$sel:healthCheckPath:EndpointGroup', endpointGroup_healthCheckPath - If the protocol is HTTP/S, then this value provides the ping path that Global Accelerator uses for the destination on the endpoints for health checks. The default is slash (/).

$sel:healthCheckIntervalSeconds:EndpointGroup', endpointGroup_healthCheckIntervalSeconds - The time—10 seconds or 30 seconds—between health checks for each endpoint. The default value is 30.

$sel:endpointGroupArn:EndpointGroup', endpointGroup_endpointGroupArn - The Amazon Resource Name (ARN) of the endpoint group.

$sel:healthCheckProtocol:EndpointGroup', endpointGroup_healthCheckProtocol - The protocol that Global Accelerator uses to perform health checks on endpoints that are part of this endpoint group. The default value is TCP.

$sel:endpointGroupRegion:EndpointGroup', endpointGroup_endpointGroupRegion - The AWS Region where the endpoint group is located.

$sel:trafficDialPercentage:EndpointGroup', endpointGroup_trafficDialPercentage - The percentage of traffic to send to an AWS Region. Additional traffic is distributed to other endpoint groups for this listener.

Use this action to increase (dial up) or decrease (dial down) traffic to a specific Region. The percentage is applied to the traffic that would otherwise have been routed to the Region based on optimal routing.

The default value is 100.

$sel:healthCheckPort:EndpointGroup', endpointGroup_healthCheckPort - The port that Global Accelerator uses to perform health checks on endpoints that are part of this endpoint group.

The default port is the port for the listener that this endpoint group is associated with. If the listener port is a list, Global Accelerator uses the first specified port in the list of ports.

$sel:portOverrides:EndpointGroup', endpointGroup_portOverrides - Allows you to override the destination ports used to route traffic to an endpoint. Using a port override lets you to map a list of external destination ports (that your users send traffic to) to a list of internal destination ports that you want an application endpoint to receive traffic on.

$sel:endpointDescriptions:EndpointGroup', endpointGroup_endpointDescriptions - The list of endpoint objects.

endpointGroup_thresholdCount :: Lens' EndpointGroup (Maybe Natural)

The number of consecutive health checks required to set the state of a healthy endpoint to unhealthy, or to set an unhealthy endpoint to healthy. The default value is 3.

endpointGroup_healthCheckPath :: Lens' EndpointGroup (Maybe Text)

If the protocol is HTTP/S, then this value provides the ping path that Global Accelerator uses for the destination on the endpoints for health checks. The default is slash (/).

endpointGroup_healthCheckIntervalSeconds :: Lens' EndpointGroup (Maybe Natural)

The time—10 seconds or 30 seconds—between health checks for each endpoint. The default value is 30.

endpointGroup_endpointGroupArn :: Lens' EndpointGroup (Maybe Text)

The Amazon Resource Name (ARN) of the endpoint group.

endpointGroup_healthCheckProtocol :: Lens' EndpointGroup (Maybe HealthCheckProtocol)

The protocol that Global Accelerator uses to perform health checks on endpoints that are part of this endpoint group. The default value is TCP.

endpointGroup_endpointGroupRegion :: Lens' EndpointGroup (Maybe Text)

The AWS Region where the endpoint group is located.

endpointGroup_trafficDialPercentage :: Lens' EndpointGroup (Maybe Double)

The percentage of traffic to send to an AWS Region. Additional traffic is distributed to other endpoint groups for this listener.

Use this action to increase (dial up) or decrease (dial down) traffic to a specific Region. The percentage is applied to the traffic that would otherwise have been routed to the Region based on optimal routing.

The default value is 100.

endpointGroup_healthCheckPort :: Lens' EndpointGroup (Maybe Natural)

The port that Global Accelerator uses to perform health checks on endpoints that are part of this endpoint group.

The default port is the port for the listener that this endpoint group is associated with. If the listener port is a list, Global Accelerator uses the first specified port in the list of ports.

endpointGroup_portOverrides :: Lens' EndpointGroup (Maybe [PortOverride])

Allows you to override the destination ports used to route traffic to an endpoint. Using a port override lets you to map a list of external destination ports (that your users send traffic to) to a list of internal destination ports that you want an application endpoint to receive traffic on.

endpointGroup_endpointDescriptions :: Lens' EndpointGroup (Maybe [EndpointDescription])

The list of endpoint objects.