Amazonka.GlobalAccelerator.UpdateAccelerator

Description

Update an accelerator.

Global Accelerator is a global service that supports endpoints in multiple AWS Regions but you must specify the US West (Oregon) Region to create or update accelerators.

Creating a Request

data UpdateAccelerator

See: newUpdateAccelerator smart constructor.

Constructors

UpdateAccelerator'

Fields

enabled :: Maybe Bool

Indicates whether an accelerator is enabled. The value is true or false. The default value is true.

If the value is set to true, the accelerator cannot be deleted. If set to false, the accelerator can be deleted.

ipAddressType :: Maybe IpAddressType

The IP address type, which must be IPv4.

name :: Maybe Text

The name of the accelerator. The name can have a maximum of 32 characters, must contain only alphanumeric characters or hyphens (-), and must not begin or end with a hyphen.

acceleratorArn :: Text

The Amazon Resource Name (ARN) of the accelerator to update.

newUpdateAccelerator

Arguments

:: Text
$sel:acceleratorArn:UpdateAccelerator'

-> UpdateAccelerator

Create a value of UpdateAccelerator with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:enabled:UpdateAccelerator', updateAccelerator_enabled - Indicates whether an accelerator is enabled. The value is true or false. The default value is true.

If the value is set to true, the accelerator cannot be deleted. If set to false, the accelerator can be deleted.

$sel:ipAddressType:UpdateAccelerator', updateAccelerator_ipAddressType - The IP address type, which must be IPv4.

$sel:name:UpdateAccelerator', updateAccelerator_name - The name of the accelerator. The name can have a maximum of 32 characters, must contain only alphanumeric characters or hyphens (-), and must not begin or end with a hyphen.

$sel:acceleratorArn:UpdateAccelerator', updateAccelerator_acceleratorArn - The Amazon Resource Name (ARN) of the accelerator to update.

Request Lenses

updateAccelerator_enabled :: Lens' UpdateAccelerator (Maybe Bool)

Indicates whether an accelerator is enabled. The value is true or false. The default value is true.

If the value is set to true, the accelerator cannot be deleted. If set to false, the accelerator can be deleted.

updateAccelerator_ipAddressType :: Lens' UpdateAccelerator (Maybe IpAddressType)

The IP address type, which must be IPv4.

updateAccelerator_name :: Lens' UpdateAccelerator (Maybe Text)

The name of the accelerator. The name can have a maximum of 32 characters, must contain only alphanumeric characters or hyphens (-), and must not begin or end with a hyphen.

updateAccelerator_acceleratorArn :: Lens' UpdateAccelerator Text

The Amazon Resource Name (ARN) of the accelerator to update.

Destructuring the Response

data UpdateAcceleratorResponse

See: newUpdateAcceleratorResponse smart constructor.

Constructors

UpdateAcceleratorResponse'

Fields

accelerator :: Maybe Accelerator

Information about the updated accelerator.

httpStatus :: Int

The response's http status code.

newUpdateAcceleratorResponse

Arguments

:: Int
$sel:httpStatus:UpdateAcceleratorResponse'

-> UpdateAcceleratorResponse

Create a value of UpdateAcceleratorResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:accelerator:UpdateAcceleratorResponse', updateAcceleratorResponse_accelerator - Information about the updated accelerator.

$sel:httpStatus:UpdateAcceleratorResponse', updateAcceleratorResponse_httpStatus - The response's http status code.

Response Lenses

updateAcceleratorResponse_accelerator :: Lens' UpdateAcceleratorResponse (Maybe Accelerator)

Information about the updated accelerator.

updateAcceleratorResponse_httpStatus :: Lens' UpdateAcceleratorResponse Int

The response's http status code.