Amazonka.GlobalAccelerator.UpdateAcceleratorAttributes

Description

Update the attributes for an accelerator.

Creating a Request

data UpdateAcceleratorAttributes

See: newUpdateAcceleratorAttributes smart constructor.

Constructors

UpdateAcceleratorAttributes'

Fields

flowLogsS3Prefix :: Maybe Text

Update the prefix for the location in the Amazon S3 bucket for the flow logs. Attribute is required if FlowLogsEnabled is true.

If you don’t specify a prefix, the flow logs are stored in the root of the bucket. If you specify slash (/) for the S3 bucket prefix, the log file bucket folder structure will include a double slash (//), like the following:

s3-bucket_name//AWSLogs/aws_account_id

flowLogsEnabled :: Maybe Bool

Update whether flow logs are enabled. The default value is false. If the value is true, FlowLogsS3Bucket and FlowLogsS3Prefix must be specified.

For more information, see Flow Logs in the AWS Global Accelerator Developer Guide.

flowLogsS3Bucket :: Maybe Text

The name of the Amazon S3 bucket for the flow logs. Attribute is required if FlowLogsEnabled is true. The bucket must exist and have a bucket policy that grants AWS Global Accelerator permission to write to the bucket.

acceleratorArn :: Text

The Amazon Resource Name (ARN) of the accelerator that you want to update.

newUpdateAcceleratorAttributes

Arguments

:: Text
$sel:acceleratorArn:UpdateAcceleratorAttributes'

-> UpdateAcceleratorAttributes

Create a value of UpdateAcceleratorAttributes with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:flowLogsS3Prefix:UpdateAcceleratorAttributes', updateAcceleratorAttributes_flowLogsS3Prefix - Update the prefix for the location in the Amazon S3 bucket for the flow logs. Attribute is required if FlowLogsEnabled is true.

If you don’t specify a prefix, the flow logs are stored in the root of the bucket. If you specify slash (/) for the S3 bucket prefix, the log file bucket folder structure will include a double slash (//), like the following:

s3-bucket_name//AWSLogs/aws_account_id

$sel:flowLogsEnabled:UpdateAcceleratorAttributes', updateAcceleratorAttributes_flowLogsEnabled - Update whether flow logs are enabled. The default value is false. If the value is true, FlowLogsS3Bucket and FlowLogsS3Prefix must be specified.

For more information, see Flow Logs in the AWS Global Accelerator Developer Guide.

$sel:flowLogsS3Bucket:UpdateAcceleratorAttributes', updateAcceleratorAttributes_flowLogsS3Bucket - The name of the Amazon S3 bucket for the flow logs. Attribute is required if FlowLogsEnabled is true. The bucket must exist and have a bucket policy that grants AWS Global Accelerator permission to write to the bucket.

$sel:acceleratorArn:UpdateAcceleratorAttributes', updateAcceleratorAttributes_acceleratorArn - The Amazon Resource Name (ARN) of the accelerator that you want to update.

Request Lenses

updateAcceleratorAttributes_flowLogsS3Prefix :: Lens' UpdateAcceleratorAttributes (Maybe Text)

Update the prefix for the location in the Amazon S3 bucket for the flow logs. Attribute is required if FlowLogsEnabled is true.

If you don’t specify a prefix, the flow logs are stored in the root of the bucket. If you specify slash (/) for the S3 bucket prefix, the log file bucket folder structure will include a double slash (//), like the following:

s3-bucket_name//AWSLogs/aws_account_id

updateAcceleratorAttributes_flowLogsEnabled :: Lens' UpdateAcceleratorAttributes (Maybe Bool)

Update whether flow logs are enabled. The default value is false. If the value is true, FlowLogsS3Bucket and FlowLogsS3Prefix must be specified.

For more information, see Flow Logs in the AWS Global Accelerator Developer Guide.

updateAcceleratorAttributes_flowLogsS3Bucket :: Lens' UpdateAcceleratorAttributes (Maybe Text)

The name of the Amazon S3 bucket for the flow logs. Attribute is required if FlowLogsEnabled is true. The bucket must exist and have a bucket policy that grants AWS Global Accelerator permission to write to the bucket.

updateAcceleratorAttributes_acceleratorArn :: Lens' UpdateAcceleratorAttributes Text

The Amazon Resource Name (ARN) of the accelerator that you want to update.

Destructuring the Response

data UpdateAcceleratorAttributesResponse

See: newUpdateAcceleratorAttributesResponse smart constructor.

Constructors

UpdateAcceleratorAttributesResponse'

Fields

acceleratorAttributes :: Maybe AcceleratorAttributes

Updated attributes for the accelerator.

httpStatus :: Int

The response's http status code.

newUpdateAcceleratorAttributesResponse

Arguments

:: Int
$sel:httpStatus:UpdateAcceleratorAttributesResponse'

-> UpdateAcceleratorAttributesResponse

Create a value of UpdateAcceleratorAttributesResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:acceleratorAttributes:UpdateAcceleratorAttributesResponse', updateAcceleratorAttributesResponse_acceleratorAttributes - Updated attributes for the accelerator.

$sel:httpStatus:UpdateAcceleratorAttributesResponse', updateAcceleratorAttributesResponse_httpStatus - The response's http status code.

Response Lenses

updateAcceleratorAttributesResponse_acceleratorAttributes :: Lens' UpdateAcceleratorAttributesResponse (Maybe AcceleratorAttributes)

Updated attributes for the accelerator.

updateAcceleratorAttributesResponse_httpStatus :: Lens' UpdateAcceleratorAttributesResponse Int

The response's http status code.