Amazonka.Glue.BatchGetWorkflows

Description

Returns a list of resource metadata for a given list of workflow names. After calling the ListWorkflows operation, you can call this operation to access the data to which you have been granted permissions. This operation supports all IAM permissions, including permission conditions that uses tags.

Creating a Request

data BatchGetWorkflows

See: newBatchGetWorkflows smart constructor.

Constructors

BatchGetWorkflows'

Fields

includeGraph :: Maybe Bool

Specifies whether to include a graph when returning the workflow resource metadata.

names :: NonEmpty Text

A list of workflow names, which may be the names returned from the ListWorkflows operation.

newBatchGetWorkflows

Arguments

:: NonEmpty Text
$sel:names:BatchGetWorkflows'

-> BatchGetWorkflows

Create a value of BatchGetWorkflows with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:includeGraph:BatchGetWorkflows', batchGetWorkflows_includeGraph - Specifies whether to include a graph when returning the workflow resource metadata.

$sel:names:BatchGetWorkflows', batchGetWorkflows_names - A list of workflow names, which may be the names returned from the ListWorkflows operation.

Request Lenses

batchGetWorkflows_includeGraph :: Lens' BatchGetWorkflows (Maybe Bool)

Specifies whether to include a graph when returning the workflow resource metadata.

batchGetWorkflows_names :: Lens' BatchGetWorkflows (NonEmpty Text)

A list of workflow names, which may be the names returned from the ListWorkflows operation.

Destructuring the Response

data BatchGetWorkflowsResponse

See: newBatchGetWorkflowsResponse smart constructor.

Constructors

BatchGetWorkflowsResponse'

Fields

missingWorkflows :: Maybe (NonEmpty Text)

A list of names of workflows not found.

workflows :: Maybe (NonEmpty Workflow)

A list of workflow resource metadata.

httpStatus :: Int

The response's http status code.

newBatchGetWorkflowsResponse

Arguments

:: Int
$sel:httpStatus:BatchGetWorkflowsResponse'

-> BatchGetWorkflowsResponse

Create a value of BatchGetWorkflowsResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:missingWorkflows:BatchGetWorkflowsResponse', batchGetWorkflowsResponse_missingWorkflows - A list of names of workflows not found.

$sel:workflows:BatchGetWorkflowsResponse', batchGetWorkflowsResponse_workflows - A list of workflow resource metadata.

$sel:httpStatus:BatchGetWorkflowsResponse', batchGetWorkflowsResponse_httpStatus - The response's http status code.

Response Lenses

batchGetWorkflowsResponse_missingWorkflows :: Lens' BatchGetWorkflowsResponse (Maybe (NonEmpty Text))

A list of names of workflows not found.

batchGetWorkflowsResponse_workflows :: Lens' BatchGetWorkflowsResponse (Maybe (NonEmpty Workflow))

A list of workflow resource metadata.

batchGetWorkflowsResponse_httpStatus :: Lens' BatchGetWorkflowsResponse Int

The response's http status code.