Amazonka.Glue.CreateDevEndpoint

Description

Creates a new development endpoint.

Creating a Request

data CreateDevEndpoint

See: newCreateDevEndpoint smart constructor.

Constructors

CreateDevEndpoint'

Fields

numberOfWorkers :: Maybe Int

The number of workers of a defined workerType that are allocated to the development endpoint.

The maximum number of workers you can define are 299 for G.1X, and 149 for G.2X.

extraPythonLibsS3Path :: Maybe Text

The paths to one or more Python libraries in an Amazon S3 bucket that should be loaded in your DevEndpoint. Multiple values must be complete paths separated by a comma.

You can only use pure Python libraries with a DevEndpoint. Libraries that rely on C extensions, such as the pandas Python data analysis library, are not yet supported.

securityGroupIds :: Maybe [Text]

Security group IDs for the security groups to be used by the new DevEndpoint.

publicKeys :: Maybe [Text]

A list of public keys to be used by the development endpoints for authentication. The use of this attribute is preferred over a single public key because the public keys allow you to have a different private key per client.

If you previously created an endpoint with a public key, you must remove that key to be able to set a list of public keys. Call the UpdateDevEndpoint API with the public key content in the deletePublicKeys attribute, and the list of new keys in the addPublicKeys attribute.

arguments :: Maybe (HashMap Text Text)

A map of arguments used to configure the DevEndpoint.

workerType :: Maybe WorkerType

The type of predefined worker that is allocated to the development endpoint. Accepts a value of Standard, G.1X, or G.2X.

For the Standard worker type, each worker provides 4 vCPU, 16 GB of memory and a 50GB disk, and 2 executors per worker.

For the G.1X worker type, each worker maps to 1 DPU (4 vCPU, 16 GB of memory, 64 GB disk), and provides 1 executor per worker. We recommend this worker type for memory-intensive jobs.

For the G.2X worker type, each worker maps to 2 DPU (8 vCPU, 32 GB of memory, 128 GB disk), and provides 1 executor per worker. We recommend this worker type for memory-intensive jobs.

Known issue: when a development endpoint is created with the G.2X WorkerType configuration, the Spark drivers for the development endpoint will run on 4 vCPU, 16 GB of memory, and a 64 GB disk.

securityConfiguration :: Maybe Text

The name of the SecurityConfiguration structure to be used with this DevEndpoint.

publicKey :: Maybe Text

The public key to be used by this DevEndpoint for authentication. This attribute is provided for backward compatibility because the recommended attribute to use is public keys.

subnetId :: Maybe Text

The subnet ID for the new DevEndpoint to use.

glueVersion :: Maybe Text

Glue version determines the versions of Apache Spark and Python that Glue supports. The Python version indicates the version supported for running your ETL scripts on development endpoints.

For more information about the available Glue versions and corresponding Spark and Python versions, see Glue version in the developer guide.

Development endpoints that are created without specifying a Glue version default to Glue 0.9.

You can specify a version of Python support for development endpoints by using the Arguments parameter in the CreateDevEndpoint or UpdateDevEndpoint APIs. If no arguments are provided, the version defaults to Python 2.

numberOfNodes :: Maybe Int

The number of Glue Data Processing Units (DPUs) to allocate to this DevEndpoint.

extraJarsS3Path :: Maybe Text

The path to one or more Java .jar files in an S3 bucket that should be loaded in your DevEndpoint.

tags :: Maybe (HashMap Text Text)

The tags to use with this DevEndpoint. You may use tags to limit access to the DevEndpoint. For more information about tags in Glue, see Amazon Web Services Tags in Glue in the developer guide.

endpointName :: Text

The name to be assigned to the new DevEndpoint.

roleArn :: Text

The IAM role for the DevEndpoint.

newCreateDevEndpoint

Arguments

:: Text
$sel:endpointName:CreateDevEndpoint'

-> Text
$sel:roleArn:CreateDevEndpoint'

-> CreateDevEndpoint

Create a value of CreateDevEndpoint with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:numberOfWorkers:CreateDevEndpoint', createDevEndpoint_numberOfWorkers - The number of workers of a defined workerType that are allocated to the development endpoint.

The maximum number of workers you can define are 299 for G.1X, and 149 for G.2X.

$sel:extraPythonLibsS3Path:CreateDevEndpoint', createDevEndpoint_extraPythonLibsS3Path - The paths to one or more Python libraries in an Amazon S3 bucket that should be loaded in your DevEndpoint. Multiple values must be complete paths separated by a comma.

You can only use pure Python libraries with a DevEndpoint. Libraries that rely on C extensions, such as the pandas Python data analysis library, are not yet supported.

$sel:securityGroupIds:CreateDevEndpoint', createDevEndpoint_securityGroupIds - Security group IDs for the security groups to be used by the new DevEndpoint.

$sel:publicKeys:CreateDevEndpoint', createDevEndpoint_publicKeys - A list of public keys to be used by the development endpoints for authentication. The use of this attribute is preferred over a single public key because the public keys allow you to have a different private key per client.

If you previously created an endpoint with a public key, you must remove that key to be able to set a list of public keys. Call the UpdateDevEndpoint API with the public key content in the deletePublicKeys attribute, and the list of new keys in the addPublicKeys attribute.

$sel:arguments:CreateDevEndpoint', createDevEndpoint_arguments - A map of arguments used to configure the DevEndpoint.

$sel:workerType:CreateDevEndpoint', createDevEndpoint_workerType - The type of predefined worker that is allocated to the development endpoint. Accepts a value of Standard, G.1X, or G.2X.

For the Standard worker type, each worker provides 4 vCPU, 16 GB of memory and a 50GB disk, and 2 executors per worker.

For the G.1X worker type, each worker maps to 1 DPU (4 vCPU, 16 GB of memory, 64 GB disk), and provides 1 executor per worker. We recommend this worker type for memory-intensive jobs.

For the G.2X worker type, each worker maps to 2 DPU (8 vCPU, 32 GB of memory, 128 GB disk), and provides 1 executor per worker. We recommend this worker type for memory-intensive jobs.

Known issue: when a development endpoint is created with the G.2X WorkerType configuration, the Spark drivers for the development endpoint will run on 4 vCPU, 16 GB of memory, and a 64 GB disk.

$sel:securityConfiguration:CreateDevEndpoint', createDevEndpoint_securityConfiguration - The name of the SecurityConfiguration structure to be used with this DevEndpoint.

$sel:publicKey:CreateDevEndpoint', createDevEndpoint_publicKey - The public key to be used by this DevEndpoint for authentication. This attribute is provided for backward compatibility because the recommended attribute to use is public keys.

$sel:subnetId:CreateDevEndpoint', createDevEndpoint_subnetId - The subnet ID for the new DevEndpoint to use.

$sel:glueVersion:CreateDevEndpoint', createDevEndpoint_glueVersion - Glue version determines the versions of Apache Spark and Python that Glue supports. The Python version indicates the version supported for running your ETL scripts on development endpoints.

For more information about the available Glue versions and corresponding Spark and Python versions, see Glue version in the developer guide.

Development endpoints that are created without specifying a Glue version default to Glue 0.9.

You can specify a version of Python support for development endpoints by using the Arguments parameter in the CreateDevEndpoint or UpdateDevEndpoint APIs. If no arguments are provided, the version defaults to Python 2.

$sel:numberOfNodes:CreateDevEndpoint', createDevEndpoint_numberOfNodes - The number of Glue Data Processing Units (DPUs) to allocate to this DevEndpoint.

$sel:extraJarsS3Path:CreateDevEndpoint', createDevEndpoint_extraJarsS3Path - The path to one or more Java .jar files in an S3 bucket that should be loaded in your DevEndpoint.

$sel:tags:CreateDevEndpoint', createDevEndpoint_tags - The tags to use with this DevEndpoint. You may use tags to limit access to the DevEndpoint. For more information about tags in Glue, see Amazon Web Services Tags in Glue in the developer guide.

$sel:endpointName:CreateDevEndpoint', createDevEndpoint_endpointName - The name to be assigned to the new DevEndpoint.

$sel:roleArn:CreateDevEndpoint', createDevEndpoint_roleArn - The IAM role for the DevEndpoint.

Request Lenses

createDevEndpoint_numberOfWorkers :: Lens' CreateDevEndpoint (Maybe Int)

The number of workers of a defined workerType that are allocated to the development endpoint.

The maximum number of workers you can define are 299 for G.1X, and 149 for G.2X.

createDevEndpoint_extraPythonLibsS3Path :: Lens' CreateDevEndpoint (Maybe Text)

The paths to one or more Python libraries in an Amazon S3 bucket that should be loaded in your DevEndpoint. Multiple values must be complete paths separated by a comma.

You can only use pure Python libraries with a DevEndpoint. Libraries that rely on C extensions, such as the pandas Python data analysis library, are not yet supported.

createDevEndpoint_securityGroupIds :: Lens' CreateDevEndpoint (Maybe [Text])

Security group IDs for the security groups to be used by the new DevEndpoint.

createDevEndpoint_publicKeys :: Lens' CreateDevEndpoint (Maybe [Text])

A list of public keys to be used by the development endpoints for authentication. The use of this attribute is preferred over a single public key because the public keys allow you to have a different private key per client.

If you previously created an endpoint with a public key, you must remove that key to be able to set a list of public keys. Call the UpdateDevEndpoint API with the public key content in the deletePublicKeys attribute, and the list of new keys in the addPublicKeys attribute.

createDevEndpoint_arguments :: Lens' CreateDevEndpoint (Maybe (HashMap Text Text))

A map of arguments used to configure the DevEndpoint.

createDevEndpoint_workerType :: Lens' CreateDevEndpoint (Maybe WorkerType)

The type of predefined worker that is allocated to the development endpoint. Accepts a value of Standard, G.1X, or G.2X.

For the Standard worker type, each worker provides 4 vCPU, 16 GB of memory and a 50GB disk, and 2 executors per worker.

For the G.1X worker type, each worker maps to 1 DPU (4 vCPU, 16 GB of memory, 64 GB disk), and provides 1 executor per worker. We recommend this worker type for memory-intensive jobs.

For the G.2X worker type, each worker maps to 2 DPU (8 vCPU, 32 GB of memory, 128 GB disk), and provides 1 executor per worker. We recommend this worker type for memory-intensive jobs.

Known issue: when a development endpoint is created with the G.2X WorkerType configuration, the Spark drivers for the development endpoint will run on 4 vCPU, 16 GB of memory, and a 64 GB disk.

createDevEndpoint_securityConfiguration :: Lens' CreateDevEndpoint (Maybe Text)

The name of the SecurityConfiguration structure to be used with this DevEndpoint.

createDevEndpoint_publicKey :: Lens' CreateDevEndpoint (Maybe Text)

The public key to be used by this DevEndpoint for authentication. This attribute is provided for backward compatibility because the recommended attribute to use is public keys.

createDevEndpoint_subnetId :: Lens' CreateDevEndpoint (Maybe Text)

The subnet ID for the new DevEndpoint to use.

createDevEndpoint_glueVersion :: Lens' CreateDevEndpoint (Maybe Text)

Glue version determines the versions of Apache Spark and Python that Glue supports. The Python version indicates the version supported for running your ETL scripts on development endpoints.

For more information about the available Glue versions and corresponding Spark and Python versions, see Glue version in the developer guide.

Development endpoints that are created without specifying a Glue version default to Glue 0.9.

You can specify a version of Python support for development endpoints by using the Arguments parameter in the CreateDevEndpoint or UpdateDevEndpoint APIs. If no arguments are provided, the version defaults to Python 2.

createDevEndpoint_numberOfNodes :: Lens' CreateDevEndpoint (Maybe Int)

The number of Glue Data Processing Units (DPUs) to allocate to this DevEndpoint.

createDevEndpoint_extraJarsS3Path :: Lens' CreateDevEndpoint (Maybe Text)

The path to one or more Java .jar files in an S3 bucket that should be loaded in your DevEndpoint.

createDevEndpoint_tags :: Lens' CreateDevEndpoint (Maybe (HashMap Text Text))

The tags to use with this DevEndpoint. You may use tags to limit access to the DevEndpoint. For more information about tags in Glue, see Amazon Web Services Tags in Glue in the developer guide.

createDevEndpoint_endpointName :: Lens' CreateDevEndpoint Text

The name to be assigned to the new DevEndpoint.

createDevEndpoint_roleArn :: Lens' CreateDevEndpoint Text

The IAM role for the DevEndpoint.

Destructuring the Response

data CreateDevEndpointResponse

See: newCreateDevEndpointResponse smart constructor.

Constructors

CreateDevEndpointResponse'

Fields

status :: Maybe Text

The current status of the new DevEndpoint.

failureReason :: Maybe Text

The reason for a current failure in this DevEndpoint.

endpointName :: Maybe Text

The name assigned to the new DevEndpoint.

numberOfWorkers :: Maybe Int

The number of workers of a defined workerType that are allocated to the development endpoint.

extraPythonLibsS3Path :: Maybe Text

The paths to one or more Python libraries in an S3 bucket that will be loaded in your DevEndpoint.

securityGroupIds :: Maybe [Text]

The security groups assigned to the new DevEndpoint.

vpcId :: Maybe Text

The ID of the virtual private cloud (VPC) used by this DevEndpoint.

arguments :: Maybe (HashMap Text Text)

The map of arguments used to configure this DevEndpoint.

Valid arguments are:

"--enable-glue-datacatalog": ""

You can specify a version of Python support for development endpoints by using the Arguments parameter in the CreateDevEndpoint or UpdateDevEndpoint APIs. If no arguments are provided, the version defaults to Python 2.

workerType :: Maybe WorkerType

The type of predefined worker that is allocated to the development endpoint. May be a value of Standard, G.1X, or G.2X.

securityConfiguration :: Maybe Text

The name of the SecurityConfiguration structure being used with this DevEndpoint.

subnetId :: Maybe Text

The subnet ID assigned to the new DevEndpoint.

glueVersion :: Maybe Text

Glue version determines the versions of Apache Spark and Python that Glue supports. The Python version indicates the version supported for running your ETL scripts on development endpoints.

For more information about the available Glue versions and corresponding Spark and Python versions, see Glue version in the developer guide.

numberOfNodes :: Maybe Int

The number of Glue Data Processing Units (DPUs) allocated to this DevEndpoint.

availabilityZone :: Maybe Text

The AWS Availability Zone where this DevEndpoint is located.

zeppelinRemoteSparkInterpreterPort :: Maybe Int

The Apache Zeppelin port for the remote Apache Spark interpreter.

extraJarsS3Path :: Maybe Text

Path to one or more Java .jar files in an S3 bucket that will be loaded in your DevEndpoint.

createdTimestamp :: Maybe POSIX

The point in time at which this DevEndpoint was created.

yarnEndpointAddress :: Maybe Text

The address of the YARN endpoint used by this DevEndpoint.

roleArn :: Maybe Text

The Amazon Resource Name (ARN) of the role assigned to the new DevEndpoint.

httpStatus :: Int

The response's http status code.

newCreateDevEndpointResponse

Arguments

:: Int
$sel:httpStatus:CreateDevEndpointResponse'

-> CreateDevEndpointResponse

Create a value of CreateDevEndpointResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:status:CreateDevEndpointResponse', createDevEndpointResponse_status - The current status of the new DevEndpoint.

$sel:failureReason:CreateDevEndpointResponse', createDevEndpointResponse_failureReason - The reason for a current failure in this DevEndpoint.

$sel:endpointName:CreateDevEndpoint', createDevEndpointResponse_endpointName - The name assigned to the new DevEndpoint.

$sel:numberOfWorkers:CreateDevEndpoint', createDevEndpointResponse_numberOfWorkers - The number of workers of a defined workerType that are allocated to the development endpoint.

$sel:extraPythonLibsS3Path:CreateDevEndpoint', createDevEndpointResponse_extraPythonLibsS3Path - The paths to one or more Python libraries in an S3 bucket that will be loaded in your DevEndpoint.

$sel:securityGroupIds:CreateDevEndpoint', createDevEndpointResponse_securityGroupIds - The security groups assigned to the new DevEndpoint.

$sel:vpcId:CreateDevEndpointResponse', createDevEndpointResponse_vpcId - The ID of the virtual private cloud (VPC) used by this DevEndpoint.

$sel:arguments:CreateDevEndpoint', createDevEndpointResponse_arguments - The map of arguments used to configure this DevEndpoint.

Valid arguments are:

"--enable-glue-datacatalog": ""

You can specify a version of Python support for development endpoints by using the Arguments parameter in the CreateDevEndpoint or UpdateDevEndpoint APIs. If no arguments are provided, the version defaults to Python 2.

$sel:workerType:CreateDevEndpoint', createDevEndpointResponse_workerType - The type of predefined worker that is allocated to the development endpoint. May be a value of Standard, G.1X, or G.2X.

$sel:securityConfiguration:CreateDevEndpoint', createDevEndpointResponse_securityConfiguration - The name of the SecurityConfiguration structure being used with this DevEndpoint.

$sel:subnetId:CreateDevEndpoint', createDevEndpointResponse_subnetId - The subnet ID assigned to the new DevEndpoint.

$sel:glueVersion:CreateDevEndpoint', createDevEndpointResponse_glueVersion - Glue version determines the versions of Apache Spark and Python that Glue supports. The Python version indicates the version supported for running your ETL scripts on development endpoints.

For more information about the available Glue versions and corresponding Spark and Python versions, see Glue version in the developer guide.

$sel:numberOfNodes:CreateDevEndpoint', createDevEndpointResponse_numberOfNodes - The number of Glue Data Processing Units (DPUs) allocated to this DevEndpoint.

$sel:availabilityZone:CreateDevEndpointResponse', createDevEndpointResponse_availabilityZone - The AWS Availability Zone where this DevEndpoint is located.

$sel:zeppelinRemoteSparkInterpreterPort:CreateDevEndpointResponse', createDevEndpointResponse_zeppelinRemoteSparkInterpreterPort - The Apache Zeppelin port for the remote Apache Spark interpreter.

$sel:extraJarsS3Path:CreateDevEndpoint', createDevEndpointResponse_extraJarsS3Path - Path to one or more Java .jar files in an S3 bucket that will be loaded in your DevEndpoint.

$sel:createdTimestamp:CreateDevEndpointResponse', createDevEndpointResponse_createdTimestamp - The point in time at which this DevEndpoint was created.

$sel:yarnEndpointAddress:CreateDevEndpointResponse', createDevEndpointResponse_yarnEndpointAddress - The address of the YARN endpoint used by this DevEndpoint.

$sel:roleArn:CreateDevEndpoint', createDevEndpointResponse_roleArn - The Amazon Resource Name (ARN) of the role assigned to the new DevEndpoint.

$sel:httpStatus:CreateDevEndpointResponse', createDevEndpointResponse_httpStatus - The response's http status code.

Response Lenses

createDevEndpointResponse_status :: Lens' CreateDevEndpointResponse (Maybe Text)

The current status of the new DevEndpoint.

createDevEndpointResponse_failureReason :: Lens' CreateDevEndpointResponse (Maybe Text)

The reason for a current failure in this DevEndpoint.

createDevEndpointResponse_endpointName :: Lens' CreateDevEndpointResponse (Maybe Text)

The name assigned to the new DevEndpoint.

createDevEndpointResponse_numberOfWorkers :: Lens' CreateDevEndpointResponse (Maybe Int)

The number of workers of a defined workerType that are allocated to the development endpoint.

createDevEndpointResponse_extraPythonLibsS3Path :: Lens' CreateDevEndpointResponse (Maybe Text)

The paths to one or more Python libraries in an S3 bucket that will be loaded in your DevEndpoint.

createDevEndpointResponse_securityGroupIds :: Lens' CreateDevEndpointResponse (Maybe [Text])

The security groups assigned to the new DevEndpoint.

createDevEndpointResponse_vpcId :: Lens' CreateDevEndpointResponse (Maybe Text)

The ID of the virtual private cloud (VPC) used by this DevEndpoint.

createDevEndpointResponse_arguments :: Lens' CreateDevEndpointResponse (Maybe (HashMap Text Text))

The map of arguments used to configure this DevEndpoint.

Valid arguments are:

"--enable-glue-datacatalog": ""

You can specify a version of Python support for development endpoints by using the Arguments parameter in the CreateDevEndpoint or UpdateDevEndpoint APIs. If no arguments are provided, the version defaults to Python 2.

createDevEndpointResponse_workerType :: Lens' CreateDevEndpointResponse (Maybe WorkerType)

The type of predefined worker that is allocated to the development endpoint. May be a value of Standard, G.1X, or G.2X.

createDevEndpointResponse_securityConfiguration :: Lens' CreateDevEndpointResponse (Maybe Text)

The name of the SecurityConfiguration structure being used with this DevEndpoint.

createDevEndpointResponse_subnetId :: Lens' CreateDevEndpointResponse (Maybe Text)

The subnet ID assigned to the new DevEndpoint.

createDevEndpointResponse_glueVersion :: Lens' CreateDevEndpointResponse (Maybe Text)

Glue version determines the versions of Apache Spark and Python that Glue supports. The Python version indicates the version supported for running your ETL scripts on development endpoints.

For more information about the available Glue versions and corresponding Spark and Python versions, see Glue version in the developer guide.

createDevEndpointResponse_numberOfNodes :: Lens' CreateDevEndpointResponse (Maybe Int)

The number of Glue Data Processing Units (DPUs) allocated to this DevEndpoint.

createDevEndpointResponse_availabilityZone :: Lens' CreateDevEndpointResponse (Maybe Text)

The AWS Availability Zone where this DevEndpoint is located.

createDevEndpointResponse_zeppelinRemoteSparkInterpreterPort :: Lens' CreateDevEndpointResponse (Maybe Int)

The Apache Zeppelin port for the remote Apache Spark interpreter.

createDevEndpointResponse_extraJarsS3Path :: Lens' CreateDevEndpointResponse (Maybe Text)

Path to one or more Java .jar files in an S3 bucket that will be loaded in your DevEndpoint.

createDevEndpointResponse_createdTimestamp :: Lens' CreateDevEndpointResponse (Maybe UTCTime)

The point in time at which this DevEndpoint was created.

createDevEndpointResponse_yarnEndpointAddress :: Lens' CreateDevEndpointResponse (Maybe Text)

The address of the YARN endpoint used by this DevEndpoint.

createDevEndpointResponse_roleArn :: Lens' CreateDevEndpointResponse (Maybe Text)

The Amazon Resource Name (ARN) of the role assigned to the new DevEndpoint.

createDevEndpointResponse_httpStatus :: Lens' CreateDevEndpointResponse Int

The response's http status code.