Amazonka.Glue.CreateJob

Description

Creates a new job definition.

Creating a Request

data CreateJob

See: newCreateJob smart constructor.

Constructors

CreateJob'

Fields

numberOfWorkers :: Maybe Int

The number of workers of a defined workerType that are allocated when a job runs.

The maximum number of workers you can define are 299 for G.1X, and 149 for G.2X.

notificationProperty :: Maybe NotificationProperty

Specifies configuration properties of a job notification.

connections :: Maybe ConnectionsList

The connections used for this job.

workerType :: Maybe WorkerType

The type of predefined worker that is allocated when a job runs. Accepts a value of Standard, G.1X, or G.2X.

For the Standard worker type, each worker provides 4 vCPU, 16 GB of memory and a 50GB disk, and 2 executors per worker.

For the G.1X worker type, each worker maps to 1 DPU (4 vCPU, 16 GB of memory, 64 GB disk), and provides 1 executor per worker. We recommend this worker type for memory-intensive jobs.

For the G.2X worker type, each worker maps to 2 DPU (8 vCPU, 32 GB of memory, 128 GB disk), and provides 1 executor per worker. We recommend this worker type for memory-intensive jobs.

securityConfiguration :: Maybe Text

The name of the SecurityConfiguration structure to be used with this job.

glueVersion :: Maybe Text

Glue version determines the versions of Apache Spark and Python that Glue supports. The Python version indicates the version supported for jobs of type Spark.

For more information about the available Glue versions and corresponding Spark and Python versions, see Glue version in the developer guide.

Jobs that are created without specifying a Glue version default to Glue 0.9.

nonOverridableArguments :: Maybe (HashMap Text Text)

Non-overridable arguments for this job, specified as name-value pairs.

logUri :: Maybe Text

This field is reserved for future use.

maxRetries :: Maybe Int

The maximum number of times to retry this job if it fails.

executionProperty :: Maybe ExecutionProperty

An ExecutionProperty specifying the maximum number of concurrent runs allowed for this job.

allocatedCapacity :: Maybe Int

This parameter is deprecated. Use MaxCapacity instead.

The number of Glue data processing units (DPUs) to allocate to this Job. You can allocate from 2 to 100 DPUs; the default is 10. A DPU is a relative measure of processing power that consists of 4 vCPUs of compute capacity and 16 GB of memory. For more information, see the Glue pricing page.

maxCapacity :: Maybe Double

For Glue version 1.0 or earlier jobs, using the standard worker type, the number of Glue data processing units (DPUs) that can be allocated when this job runs. A DPU is a relative measure of processing power that consists of 4 vCPUs of compute capacity and 16 GB of memory. For more information, see the Glue pricing page.

Do not set Max Capacity if using WorkerType and NumberOfWorkers.

The value that can be allocated for MaxCapacity depends on whether you are running a Python shell job or an Apache Spark ETL job:

When you specify a Python shell job (JobCommand.Name="pythonshell"), you can allocate either 0.0625 or 1 DPU. The default is 0.0625 DPU.

When you specify an Apache Spark ETL job (JobCommand.Name="glueetl") or Apache Spark streaming ETL job (JobCommand.Name="gluestreaming"), you can allocate from 2 to 100 DPUs. The default is 10 DPUs. This job type cannot have a fractional DPU allocation.

For Glue version 2.0 jobs, you cannot instead specify a Maximum capacity. Instead, you should specify a Worker type and the Number of workers.

timeout :: Maybe Natural

The job timeout in minutes. This is the maximum time that a job run can consume resources before it is terminated and enters TIMEOUT status. The default is 2,880 minutes (48 hours).

defaultArguments :: Maybe (HashMap Text Text)

The default arguments for this job.

You can specify arguments here that your own job-execution script consumes, as well as arguments that Glue itself consumes.

For information about how to specify and consume your own Job arguments, see the Calling Glue APIs in Python topic in the developer guide.

For information about the key-value pairs that Glue consumes to set up your job, see the Special Parameters Used by Glue topic in the developer guide.

description :: Maybe Text

Description of the job being defined.

tags :: Maybe (HashMap Text Text)

The tags to use with this job. You may use tags to limit access to the job. For more information about tags in Glue, see Amazon Web Services Tags in Glue in the developer guide.

name :: Text

The name you assign to this job definition. It must be unique in your account.

role' :: Text

The name or Amazon Resource Name (ARN) of the IAM role associated with this job.

command :: JobCommand

The JobCommand that runs this job.

newCreateJob

Arguments

:: Text
$sel:name:CreateJob'

-> Text
$sel:role':CreateJob'

-> JobCommand
$sel:command:CreateJob'

-> CreateJob

Create a value of CreateJob with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:numberOfWorkers:CreateJob', createJob_numberOfWorkers - The number of workers of a defined workerType that are allocated when a job runs.

The maximum number of workers you can define are 299 for G.1X, and 149 for G.2X.

$sel:notificationProperty:CreateJob', createJob_notificationProperty - Specifies configuration properties of a job notification.

$sel:connections:CreateJob', createJob_connections - The connections used for this job.

$sel:workerType:CreateJob', createJob_workerType - The type of predefined worker that is allocated when a job runs. Accepts a value of Standard, G.1X, or G.2X.

For the Standard worker type, each worker provides 4 vCPU, 16 GB of memory and a 50GB disk, and 2 executors per worker.

For the G.1X worker type, each worker maps to 1 DPU (4 vCPU, 16 GB of memory, 64 GB disk), and provides 1 executor per worker. We recommend this worker type for memory-intensive jobs.

For the G.2X worker type, each worker maps to 2 DPU (8 vCPU, 32 GB of memory, 128 GB disk), and provides 1 executor per worker. We recommend this worker type for memory-intensive jobs.

$sel:securityConfiguration:CreateJob', createJob_securityConfiguration - The name of the SecurityConfiguration structure to be used with this job.

$sel:glueVersion:CreateJob', createJob_glueVersion - Glue version determines the versions of Apache Spark and Python that Glue supports. The Python version indicates the version supported for jobs of type Spark.

For more information about the available Glue versions and corresponding Spark and Python versions, see Glue version in the developer guide.

Jobs that are created without specifying a Glue version default to Glue 0.9.

$sel:nonOverridableArguments:CreateJob', createJob_nonOverridableArguments - Non-overridable arguments for this job, specified as name-value pairs.

$sel:logUri:CreateJob', createJob_logUri - This field is reserved for future use.

$sel:maxRetries:CreateJob', createJob_maxRetries - The maximum number of times to retry this job if it fails.

$sel:executionProperty:CreateJob', createJob_executionProperty - An ExecutionProperty specifying the maximum number of concurrent runs allowed for this job.

$sel:allocatedCapacity:CreateJob', createJob_allocatedCapacity - This parameter is deprecated. Use MaxCapacity instead.

The number of Glue data processing units (DPUs) to allocate to this Job. You can allocate from 2 to 100 DPUs; the default is 10. A DPU is a relative measure of processing power that consists of 4 vCPUs of compute capacity and 16 GB of memory. For more information, see the Glue pricing page.

$sel:maxCapacity:CreateJob', createJob_maxCapacity - For Glue version 1.0 or earlier jobs, using the standard worker type, the number of Glue data processing units (DPUs) that can be allocated when this job runs. A DPU is a relative measure of processing power that consists of 4 vCPUs of compute capacity and 16 GB of memory. For more information, see the Glue pricing page.

Do not set Max Capacity if using WorkerType and NumberOfWorkers.

The value that can be allocated for MaxCapacity depends on whether you are running a Python shell job or an Apache Spark ETL job:

When you specify a Python shell job (JobCommand.Name="pythonshell"), you can allocate either 0.0625 or 1 DPU. The default is 0.0625 DPU.

When you specify an Apache Spark ETL job (JobCommand.Name="glueetl") or Apache Spark streaming ETL job (JobCommand.Name="gluestreaming"), you can allocate from 2 to 100 DPUs. The default is 10 DPUs. This job type cannot have a fractional DPU allocation.

For Glue version 2.0 jobs, you cannot instead specify a Maximum capacity. Instead, you should specify a Worker type and the Number of workers.

$sel:timeout:CreateJob', createJob_timeout - The job timeout in minutes. This is the maximum time that a job run can consume resources before it is terminated and enters TIMEOUT status. The default is 2,880 minutes (48 hours).

$sel:defaultArguments:CreateJob', createJob_defaultArguments - The default arguments for this job.

You can specify arguments here that your own job-execution script consumes, as well as arguments that Glue itself consumes.

For information about how to specify and consume your own Job arguments, see the Calling Glue APIs in Python topic in the developer guide.

For information about the key-value pairs that Glue consumes to set up your job, see the Special Parameters Used by Glue topic in the developer guide.

$sel:description:CreateJob', createJob_description - Description of the job being defined.

$sel:tags:CreateJob', createJob_tags - The tags to use with this job. You may use tags to limit access to the job. For more information about tags in Glue, see Amazon Web Services Tags in Glue in the developer guide.

$sel:name:CreateJob', createJob_name - The name you assign to this job definition. It must be unique in your account.

$sel:role':CreateJob', createJob_role - The name or Amazon Resource Name (ARN) of the IAM role associated with this job.

$sel:command:CreateJob', createJob_command - The JobCommand that runs this job.

Request Lenses

createJob_numberOfWorkers :: Lens' CreateJob (Maybe Int)

The number of workers of a defined workerType that are allocated when a job runs.

The maximum number of workers you can define are 299 for G.1X, and 149 for G.2X.

createJob_notificationProperty :: Lens' CreateJob (Maybe NotificationProperty)

Specifies configuration properties of a job notification.

createJob_connections :: Lens' CreateJob (Maybe ConnectionsList)

The connections used for this job.

createJob_workerType :: Lens' CreateJob (Maybe WorkerType)

The type of predefined worker that is allocated when a job runs. Accepts a value of Standard, G.1X, or G.2X.

For the Standard worker type, each worker provides 4 vCPU, 16 GB of memory and a 50GB disk, and 2 executors per worker.

For the G.1X worker type, each worker maps to 1 DPU (4 vCPU, 16 GB of memory, 64 GB disk), and provides 1 executor per worker. We recommend this worker type for memory-intensive jobs.

For the G.2X worker type, each worker maps to 2 DPU (8 vCPU, 32 GB of memory, 128 GB disk), and provides 1 executor per worker. We recommend this worker type for memory-intensive jobs.

createJob_securityConfiguration :: Lens' CreateJob (Maybe Text)

The name of the SecurityConfiguration structure to be used with this job.

createJob_glueVersion :: Lens' CreateJob (Maybe Text)

Glue version determines the versions of Apache Spark and Python that Glue supports. The Python version indicates the version supported for jobs of type Spark.

For more information about the available Glue versions and corresponding Spark and Python versions, see Glue version in the developer guide.

Jobs that are created without specifying a Glue version default to Glue 0.9.

createJob_nonOverridableArguments :: Lens' CreateJob (Maybe (HashMap Text Text))

Non-overridable arguments for this job, specified as name-value pairs.

createJob_logUri :: Lens' CreateJob (Maybe Text)

This field is reserved for future use.

createJob_maxRetries :: Lens' CreateJob (Maybe Int)

The maximum number of times to retry this job if it fails.

createJob_executionProperty :: Lens' CreateJob (Maybe ExecutionProperty)

An ExecutionProperty specifying the maximum number of concurrent runs allowed for this job.

createJob_allocatedCapacity :: Lens' CreateJob (Maybe Int)

This parameter is deprecated. Use MaxCapacity instead.

The number of Glue data processing units (DPUs) to allocate to this Job. You can allocate from 2 to 100 DPUs; the default is 10. A DPU is a relative measure of processing power that consists of 4 vCPUs of compute capacity and 16 GB of memory. For more information, see the Glue pricing page.

createJob_maxCapacity :: Lens' CreateJob (Maybe Double)

For Glue version 1.0 or earlier jobs, using the standard worker type, the number of Glue data processing units (DPUs) that can be allocated when this job runs. A DPU is a relative measure of processing power that consists of 4 vCPUs of compute capacity and 16 GB of memory. For more information, see the Glue pricing page.

Do not set Max Capacity if using WorkerType and NumberOfWorkers.

The value that can be allocated for MaxCapacity depends on whether you are running a Python shell job or an Apache Spark ETL job:

When you specify a Python shell job (JobCommand.Name="pythonshell"), you can allocate either 0.0625 or 1 DPU. The default is 0.0625 DPU.

When you specify an Apache Spark ETL job (JobCommand.Name="glueetl") or Apache Spark streaming ETL job (JobCommand.Name="gluestreaming"), you can allocate from 2 to 100 DPUs. The default is 10 DPUs. This job type cannot have a fractional DPU allocation.

For Glue version 2.0 jobs, you cannot instead specify a Maximum capacity. Instead, you should specify a Worker type and the Number of workers.

createJob_timeout :: Lens' CreateJob (Maybe Natural)

The job timeout in minutes. This is the maximum time that a job run can consume resources before it is terminated and enters TIMEOUT status. The default is 2,880 minutes (48 hours).

createJob_defaultArguments :: Lens' CreateJob (Maybe (HashMap Text Text))

The default arguments for this job.

You can specify arguments here that your own job-execution script consumes, as well as arguments that Glue itself consumes.

For information about how to specify and consume your own Job arguments, see the Calling Glue APIs in Python topic in the developer guide.

For information about the key-value pairs that Glue consumes to set up your job, see the Special Parameters Used by Glue topic in the developer guide.

createJob_description :: Lens' CreateJob (Maybe Text)

Description of the job being defined.

createJob_tags :: Lens' CreateJob (Maybe (HashMap Text Text))

The tags to use with this job. You may use tags to limit access to the job. For more information about tags in Glue, see Amazon Web Services Tags in Glue in the developer guide.

createJob_name :: Lens' CreateJob Text

The name you assign to this job definition. It must be unique in your account.

createJob_role :: Lens' CreateJob Text

The name or Amazon Resource Name (ARN) of the IAM role associated with this job.

createJob_command :: Lens' CreateJob JobCommand

The JobCommand that runs this job.

Destructuring the Response

data CreateJobResponse

See: newCreateJobResponse smart constructor.

Constructors

CreateJobResponse'

Fields

name :: Maybe Text

The unique name that was provided for this job definition.

httpStatus :: Int

The response's http status code.

newCreateJobResponse

Arguments

:: Int
$sel:httpStatus:CreateJobResponse'

-> CreateJobResponse

Create a value of CreateJobResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:name:CreateJob', createJobResponse_name - The unique name that was provided for this job definition.

$sel:httpStatus:CreateJobResponse', createJobResponse_httpStatus - The response's http status code.

Response Lenses

createJobResponse_name :: Lens' CreateJobResponse (Maybe Text)

The unique name that was provided for this job definition.

createJobResponse_httpStatus :: Lens' CreateJobResponse Int

The response's http status code.