Amazonka.Glue.DeleteSchemaVersions

Description

Remove versions from the specified schema. A version number or range may be supplied. If the compatibility mode forbids deleting of a version that is necessary, such as BACKWARDS_FULL, an error is returned. Calling the GetSchemaVersions API after this call will list the status of the deleted versions.

When the range of version numbers contain check pointed version, the API will return a 409 conflict and will not proceed with the deletion. You have to remove the checkpoint first using the DeleteSchemaCheckpoint API before using this API.

You cannot use the DeleteSchemaVersions API to delete the first schema version in the schema set. The first schema version can only be deleted by the DeleteSchema API. This operation will also delete the attached SchemaVersionMetadata under the schema versions. Hard deletes will be enforced on the database.

If the compatibility mode forbids deleting of a version that is necessary, such as BACKWARDS_FULL, an error is returned.

Creating a Request

data DeleteSchemaVersions

See: newDeleteSchemaVersions smart constructor.

Constructors

DeleteSchemaVersions'

Fields

schemaId :: SchemaId

This is a wrapper structure that may contain the schema name and Amazon Resource Name (ARN).

versions :: Text

A version range may be supplied which may be of the format:

a single version number, 5

a range, 5-8 : deletes versions 5, 6, 7, 8

newDeleteSchemaVersions

Arguments

:: SchemaId
$sel:schemaId:DeleteSchemaVersions'

-> Text
$sel:versions:DeleteSchemaVersions'

-> DeleteSchemaVersions

Create a value of DeleteSchemaVersions with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:schemaId:DeleteSchemaVersions', deleteSchemaVersions_schemaId - This is a wrapper structure that may contain the schema name and Amazon Resource Name (ARN).

$sel:versions:DeleteSchemaVersions', deleteSchemaVersions_versions - A version range may be supplied which may be of the format:

a single version number, 5

a range, 5-8 : deletes versions 5, 6, 7, 8

Request Lenses

deleteSchemaVersions_schemaId :: Lens' DeleteSchemaVersions SchemaId

This is a wrapper structure that may contain the schema name and Amazon Resource Name (ARN).

deleteSchemaVersions_versions :: Lens' DeleteSchemaVersions Text

A version range may be supplied which may be of the format:

a single version number, 5

a range, 5-8 : deletes versions 5, 6, 7, 8

Destructuring the Response

data DeleteSchemaVersionsResponse

See: newDeleteSchemaVersionsResponse smart constructor.

Constructors

DeleteSchemaVersionsResponse'

Fields

schemaVersionErrors :: Maybe [SchemaVersionErrorItem]

A list of SchemaVersionErrorItem objects, each containing an error and schema version.

httpStatus :: Int

The response's http status code.

newDeleteSchemaVersionsResponse

Arguments

:: Int
$sel:httpStatus:DeleteSchemaVersionsResponse'

-> DeleteSchemaVersionsResponse

Create a value of DeleteSchemaVersionsResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:schemaVersionErrors:DeleteSchemaVersionsResponse', deleteSchemaVersionsResponse_schemaVersionErrors - A list of SchemaVersionErrorItem objects, each containing an error and schema version.

$sel:httpStatus:DeleteSchemaVersionsResponse', deleteSchemaVersionsResponse_httpStatus - The response's http status code.

Response Lenses

deleteSchemaVersionsResponse_schemaVersionErrors :: Lens' DeleteSchemaVersionsResponse (Maybe [SchemaVersionErrorItem])

A list of SchemaVersionErrorItem objects, each containing an error and schema version.

deleteSchemaVersionsResponse_httpStatus :: Lens' DeleteSchemaVersionsResponse Int

The response's http status code.