Amazonka.Glue.GetColumnStatisticsForTable

Description

Retrieves table statistics of columns.

The Identity and Access Management (IAM) permission required for this operation is GetTable.

Creating a Request

data GetColumnStatisticsForTable

See: newGetColumnStatisticsForTable smart constructor.

Constructors

GetColumnStatisticsForTable'

Fields

catalogId :: Maybe Text

The ID of the Data Catalog where the partitions in question reside. If none is supplied, the Amazon Web Services account ID is used by default.

databaseName :: Text

The name of the catalog database where the partitions reside.

tableName :: Text

The name of the partitions' table.

columnNames :: [Text]

A list of the column names.

newGetColumnStatisticsForTable

Arguments

:: Text
$sel:databaseName:GetColumnStatisticsForTable'

-> Text
$sel:tableName:GetColumnStatisticsForTable'

-> GetColumnStatisticsForTable

Create a value of GetColumnStatisticsForTable with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:catalogId:GetColumnStatisticsForTable', getColumnStatisticsForTable_catalogId - The ID of the Data Catalog where the partitions in question reside. If none is supplied, the Amazon Web Services account ID is used by default.

$sel:databaseName:GetColumnStatisticsForTable', getColumnStatisticsForTable_databaseName - The name of the catalog database where the partitions reside.

$sel:tableName:GetColumnStatisticsForTable', getColumnStatisticsForTable_tableName - The name of the partitions' table.

$sel:columnNames:GetColumnStatisticsForTable', getColumnStatisticsForTable_columnNames - A list of the column names.

Request Lenses

getColumnStatisticsForTable_catalogId :: Lens' GetColumnStatisticsForTable (Maybe Text)

The ID of the Data Catalog where the partitions in question reside. If none is supplied, the Amazon Web Services account ID is used by default.

getColumnStatisticsForTable_databaseName :: Lens' GetColumnStatisticsForTable Text

The name of the catalog database where the partitions reside.

getColumnStatisticsForTable_tableName :: Lens' GetColumnStatisticsForTable Text

The name of the partitions' table.

getColumnStatisticsForTable_columnNames :: Lens' GetColumnStatisticsForTable [Text]

A list of the column names.

Destructuring the Response

data GetColumnStatisticsForTableResponse

See: newGetColumnStatisticsForTableResponse smart constructor.

Constructors

GetColumnStatisticsForTableResponse'

Fields

errors :: Maybe [ColumnError]

List of ColumnStatistics that failed to be retrieved.

columnStatisticsList :: Maybe [ColumnStatistics]

List of ColumnStatistics that failed to be retrieved.

httpStatus :: Int

The response's http status code.

newGetColumnStatisticsForTableResponse

Arguments

:: Int
$sel:httpStatus:GetColumnStatisticsForTableResponse'

-> GetColumnStatisticsForTableResponse

Create a value of GetColumnStatisticsForTableResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:errors:GetColumnStatisticsForTableResponse', getColumnStatisticsForTableResponse_errors - List of ColumnStatistics that failed to be retrieved.

$sel:columnStatisticsList:GetColumnStatisticsForTableResponse', getColumnStatisticsForTableResponse_columnStatisticsList - List of ColumnStatistics that failed to be retrieved.

$sel:httpStatus:GetColumnStatisticsForTableResponse', getColumnStatisticsForTableResponse_httpStatus - The response's http status code.

Response Lenses

getColumnStatisticsForTableResponse_errors :: Lens' GetColumnStatisticsForTableResponse (Maybe [ColumnError])

List of ColumnStatistics that failed to be retrieved.

getColumnStatisticsForTableResponse_columnStatisticsList :: Lens' GetColumnStatisticsForTableResponse (Maybe [ColumnStatistics])

List of ColumnStatistics that failed to be retrieved.

getColumnStatisticsForTableResponse_httpStatus :: Lens' GetColumnStatisticsForTableResponse Int

The response's http status code.