Amazonka.Glue.GetDatabases

Description

Retrieves all databases defined in a given Data Catalog.

This operation returns paginated results.

Creating a Request

data GetDatabases

See: newGetDatabases smart constructor.

Constructors

GetDatabases'

Fields

resourceShareType :: Maybe ResourceShareType

Allows you to specify that you want to list the databases shared with your account. The allowable values are FOREIGN or ALL.

If set to FOREIGN, will list the databases shared with your account.

If set to ALL, will list the databases shared with your account, as well as the databases in yor local account.

catalogId :: Maybe Text

The ID of the Data Catalog from which to retrieve Databases. If none is provided, the Amazon Web Services account ID is used by default.

nextToken :: Maybe Text

A continuation token, if this is a continuation call.

maxResults :: Maybe Natural

The maximum number of databases to return in one response.

newGetDatabases :: GetDatabases

Create a value of GetDatabases with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:resourceShareType:GetDatabases', getDatabases_resourceShareType - Allows you to specify that you want to list the databases shared with your account. The allowable values are FOREIGN or ALL.

If set to FOREIGN, will list the databases shared with your account.

If set to ALL, will list the databases shared with your account, as well as the databases in yor local account.

$sel:catalogId:GetDatabases', getDatabases_catalogId - The ID of the Data Catalog from which to retrieve Databases. If none is provided, the Amazon Web Services account ID is used by default.

$sel:nextToken:GetDatabases', getDatabases_nextToken - A continuation token, if this is a continuation call.

$sel:maxResults:GetDatabases', getDatabases_maxResults - The maximum number of databases to return in one response.

Request Lenses

getDatabases_resourceShareType :: Lens' GetDatabases (Maybe ResourceShareType)

Allows you to specify that you want to list the databases shared with your account. The allowable values are FOREIGN or ALL.

If set to FOREIGN, will list the databases shared with your account.

If set to ALL, will list the databases shared with your account, as well as the databases in yor local account.

getDatabases_catalogId :: Lens' GetDatabases (Maybe Text)

The ID of the Data Catalog from which to retrieve Databases. If none is provided, the Amazon Web Services account ID is used by default.

getDatabases_nextToken :: Lens' GetDatabases (Maybe Text)

A continuation token, if this is a continuation call.

getDatabases_maxResults :: Lens' GetDatabases (Maybe Natural)

The maximum number of databases to return in one response.

Destructuring the Response

data GetDatabasesResponse

See: newGetDatabasesResponse smart constructor.

Constructors

GetDatabasesResponse'

Fields

nextToken :: Maybe Text

A continuation token for paginating the returned list of tokens, returned if the current segment of the list is not the last.

httpStatus :: Int

The response's http status code.

databaseList :: [Database]

A list of Database objects from the specified catalog.

newGetDatabasesResponse

Arguments

:: Int
$sel:httpStatus:GetDatabasesResponse'

-> GetDatabasesResponse

Create a value of GetDatabasesResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:nextToken:GetDatabases', getDatabasesResponse_nextToken - A continuation token for paginating the returned list of tokens, returned if the current segment of the list is not the last.

$sel:httpStatus:GetDatabasesResponse', getDatabasesResponse_httpStatus - The response's http status code.

$sel:databaseList:GetDatabasesResponse', getDatabasesResponse_databaseList - A list of Database objects from the specified catalog.

Response Lenses

getDatabasesResponse_nextToken :: Lens' GetDatabasesResponse (Maybe Text)

A continuation token for paginating the returned list of tokens, returned if the current segment of the list is not the last.

getDatabasesResponse_httpStatus :: Lens' GetDatabasesResponse Int

The response's http status code.

getDatabasesResponse_databaseList :: Lens' GetDatabasesResponse [Database]

A list of Database objects from the specified catalog.