Amazonka.Glue.GetJobRuns

Description

Retrieves metadata for all runs of a given job definition.

This operation returns paginated results.

Creating a Request

data GetJobRuns

See: newGetJobRuns smart constructor.

Constructors

GetJobRuns'

Fields

nextToken :: Maybe Text

A continuation token, if this is a continuation call.

maxResults :: Maybe Natural

The maximum size of the response.

jobName :: Text

The name of the job definition for which to retrieve all job runs.

newGetJobRuns

Arguments

:: Text
$sel:jobName:GetJobRuns'

-> GetJobRuns

Create a value of GetJobRuns with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:nextToken:GetJobRuns', getJobRuns_nextToken - A continuation token, if this is a continuation call.

$sel:maxResults:GetJobRuns', getJobRuns_maxResults - The maximum size of the response.

$sel:jobName:GetJobRuns', getJobRuns_jobName - The name of the job definition for which to retrieve all job runs.

Request Lenses

getJobRuns_nextToken :: Lens' GetJobRuns (Maybe Text)

A continuation token, if this is a continuation call.

getJobRuns_maxResults :: Lens' GetJobRuns (Maybe Natural)

The maximum size of the response.

getJobRuns_jobName :: Lens' GetJobRuns Text

The name of the job definition for which to retrieve all job runs.

Destructuring the Response

data GetJobRunsResponse

See: newGetJobRunsResponse smart constructor.

Constructors

GetJobRunsResponse'

Fields

nextToken :: Maybe Text

A continuation token, if not all requested job runs have been returned.

jobRuns :: Maybe [JobRun]

A list of job-run metadata objects.

httpStatus :: Int

The response's http status code.

newGetJobRunsResponse

Arguments

:: Int
$sel:httpStatus:GetJobRunsResponse'

-> GetJobRunsResponse

Create a value of GetJobRunsResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:nextToken:GetJobRuns', getJobRunsResponse_nextToken - A continuation token, if not all requested job runs have been returned.

$sel:jobRuns:GetJobRunsResponse', getJobRunsResponse_jobRuns - A list of job-run metadata objects.

$sel:httpStatus:GetJobRunsResponse', getJobRunsResponse_httpStatus - The response's http status code.

Response Lenses

getJobRunsResponse_nextToken :: Lens' GetJobRunsResponse (Maybe Text)

A continuation token, if not all requested job runs have been returned.

getJobRunsResponse_jobRuns :: Lens' GetJobRunsResponse (Maybe [JobRun])

A list of job-run metadata objects.

getJobRunsResponse_httpStatus :: Lens' GetJobRunsResponse Int

The response's http status code.