Amazonka.Glue.GetMLTransform

Description

Gets an Glue machine learning transform artifact and all its corresponding metadata. Machine learning transforms are a special type of transform that use machine learning to learn the details of the transformation to be performed by learning from examples provided by humans. These transformations are then saved by Glue. You can retrieve their metadata by calling GetMLTransform.

Creating a Request

data GetMLTransform

See: newGetMLTransform smart constructor.

Constructors

GetMLTransform'

Fields

transformId :: Text

The unique identifier of the transform, generated at the time that the transform was created.

newGetMLTransform

Arguments

:: Text
$sel:transformId:GetMLTransform'

-> GetMLTransform

Create a value of GetMLTransform with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:transformId:GetMLTransform', getMLTransform_transformId - The unique identifier of the transform, generated at the time that the transform was created.

Request Lenses

getMLTransform_transformId :: Lens' GetMLTransform Text

The unique identifier of the transform, generated at the time that the transform was created.

Destructuring the Response

data GetMLTransformResponse

See: newGetMLTransformResponse smart constructor.

Constructors

GetMLTransformResponse'

Fields

status :: Maybe TransformStatusType

The last known status of the transform (to indicate whether it can be used or not). One of "NOT_READY", "READY", or "DELETING".

numberOfWorkers :: Maybe Int

The number of workers of a defined workerType that are allocated when this task runs.

lastModifiedOn :: Maybe POSIX

The date and time when the transform was last modified.

labelCount :: Maybe Int

The number of labels available for this transform.

workerType :: Maybe WorkerType

The type of predefined worker that is allocated when this task runs. Accepts a value of Standard, G.1X, or G.2X.

For the Standard worker type, each worker provides 4 vCPU, 16 GB of memory and a 50GB disk, and 2 executors per worker.

For the G.1X worker type, each worker provides 4 vCPU, 16 GB of memory and a 64GB disk, and 1 executor per worker.

For the G.2X worker type, each worker provides 8 vCPU, 32 GB of memory and a 128GB disk, and 1 executor per worker.

inputRecordTables :: Maybe [GlueTable]

A list of Glue table definitions used by the transform.

glueVersion :: Maybe Text

This value determines which version of Glue this machine learning transform is compatible with. Glue 1.0 is recommended for most customers. If the value is not set, the Glue compatibility defaults to Glue 0.9. For more information, see Glue Versions in the developer guide.

evaluationMetrics :: Maybe EvaluationMetrics

The latest evaluation metrics.

schema :: Maybe [SchemaColumn]

The Map<Column, Type> object that represents the schema that this transform accepts. Has an upper bound of 100 columns.

role' :: Maybe Text

The name or Amazon Resource Name (ARN) of the IAM role with the required permissions.

name :: Maybe Text

The unique name given to the transform when it was created.

parameters :: Maybe TransformParameters

The configuration parameters that are specific to the algorithm used.

maxRetries :: Maybe Int

The maximum number of times to retry a task for this transform after a task run fails.

maxCapacity :: Maybe Double

The number of Glue data processing units (DPUs) that are allocated to task runs for this transform. You can allocate from 2 to 100 DPUs; the default is 10. A DPU is a relative measure of processing power that consists of 4 vCPUs of compute capacity and 16 GB of memory. For more information, see the Glue pricing page.

When the WorkerType field is set to a value other than Standard, the MaxCapacity field is set automatically and becomes read-only.

timeout :: Maybe Natural

The timeout for a task run for this transform in minutes. This is the maximum time that a task run for this transform can consume resources before it is terminated and enters TIMEOUT status. The default is 2,880 minutes (48 hours).

transformEncryption :: Maybe TransformEncryption

The encryption-at-rest settings of the transform that apply to accessing user data. Machine learning transforms can access user data encrypted in Amazon S3 using KMS.

description :: Maybe Text

A description of the transform.

createdOn :: Maybe POSIX

The date and time when the transform was created.

transformId :: Maybe Text

The unique identifier of the transform, generated at the time that the transform was created.

httpStatus :: Int

The response's http status code.

newGetMLTransformResponse

Arguments

:: Int
$sel:httpStatus:GetMLTransformResponse'

-> GetMLTransformResponse

Create a value of GetMLTransformResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:status:GetMLTransformResponse', getMLTransformResponse_status - The last known status of the transform (to indicate whether it can be used or not). One of "NOT_READY", "READY", or "DELETING".

$sel:numberOfWorkers:GetMLTransformResponse', getMLTransformResponse_numberOfWorkers - The number of workers of a defined workerType that are allocated when this task runs.

$sel:lastModifiedOn:GetMLTransformResponse', getMLTransformResponse_lastModifiedOn - The date and time when the transform was last modified.

$sel:labelCount:GetMLTransformResponse', getMLTransformResponse_labelCount - The number of labels available for this transform.

$sel:workerType:GetMLTransformResponse', getMLTransformResponse_workerType - The type of predefined worker that is allocated when this task runs. Accepts a value of Standard, G.1X, or G.2X.

For the Standard worker type, each worker provides 4 vCPU, 16 GB of memory and a 50GB disk, and 2 executors per worker.

For the G.1X worker type, each worker provides 4 vCPU, 16 GB of memory and a 64GB disk, and 1 executor per worker.

For the G.2X worker type, each worker provides 8 vCPU, 32 GB of memory and a 128GB disk, and 1 executor per worker.

$sel:inputRecordTables:GetMLTransformResponse', getMLTransformResponse_inputRecordTables - A list of Glue table definitions used by the transform.

$sel:glueVersion:GetMLTransformResponse', getMLTransformResponse_glueVersion - This value determines which version of Glue this machine learning transform is compatible with. Glue 1.0 is recommended for most customers. If the value is not set, the Glue compatibility defaults to Glue 0.9. For more information, see Glue Versions in the developer guide.

$sel:evaluationMetrics:GetMLTransformResponse', getMLTransformResponse_evaluationMetrics - The latest evaluation metrics.

$sel:schema:GetMLTransformResponse', getMLTransformResponse_schema - The Map<Column, Type> object that represents the schema that this transform accepts. Has an upper bound of 100 columns.

$sel:role':GetMLTransformResponse', getMLTransformResponse_role - The name or Amazon Resource Name (ARN) of the IAM role with the required permissions.

$sel:name:GetMLTransformResponse', getMLTransformResponse_name - The unique name given to the transform when it was created.

$sel:parameters:GetMLTransformResponse', getMLTransformResponse_parameters - The configuration parameters that are specific to the algorithm used.

$sel:maxRetries:GetMLTransformResponse', getMLTransformResponse_maxRetries - The maximum number of times to retry a task for this transform after a task run fails.

$sel:maxCapacity:GetMLTransformResponse', getMLTransformResponse_maxCapacity - The number of Glue data processing units (DPUs) that are allocated to task runs for this transform. You can allocate from 2 to 100 DPUs; the default is 10. A DPU is a relative measure of processing power that consists of 4 vCPUs of compute capacity and 16 GB of memory. For more information, see the Glue pricing page.

When the WorkerType field is set to a value other than Standard, the MaxCapacity field is set automatically and becomes read-only.

$sel:timeout:GetMLTransformResponse', getMLTransformResponse_timeout - The timeout for a task run for this transform in minutes. This is the maximum time that a task run for this transform can consume resources before it is terminated and enters TIMEOUT status. The default is 2,880 minutes (48 hours).

$sel:transformEncryption:GetMLTransformResponse', getMLTransformResponse_transformEncryption - The encryption-at-rest settings of the transform that apply to accessing user data. Machine learning transforms can access user data encrypted in Amazon S3 using KMS.

$sel:description:GetMLTransformResponse', getMLTransformResponse_description - A description of the transform.

$sel:createdOn:GetMLTransformResponse', getMLTransformResponse_createdOn - The date and time when the transform was created.

$sel:transformId:GetMLTransform', getMLTransformResponse_transformId - The unique identifier of the transform, generated at the time that the transform was created.

$sel:httpStatus:GetMLTransformResponse', getMLTransformResponse_httpStatus - The response's http status code.

Response Lenses

getMLTransformResponse_status :: Lens' GetMLTransformResponse (Maybe TransformStatusType)

The last known status of the transform (to indicate whether it can be used or not). One of "NOT_READY", "READY", or "DELETING".

getMLTransformResponse_numberOfWorkers :: Lens' GetMLTransformResponse (Maybe Int)

The number of workers of a defined workerType that are allocated when this task runs.

getMLTransformResponse_lastModifiedOn :: Lens' GetMLTransformResponse (Maybe UTCTime)

The date and time when the transform was last modified.

getMLTransformResponse_labelCount :: Lens' GetMLTransformResponse (Maybe Int)

The number of labels available for this transform.

getMLTransformResponse_workerType :: Lens' GetMLTransformResponse (Maybe WorkerType)

The type of predefined worker that is allocated when this task runs. Accepts a value of Standard, G.1X, or G.2X.

For the Standard worker type, each worker provides 4 vCPU, 16 GB of memory and a 50GB disk, and 2 executors per worker.

For the G.1X worker type, each worker provides 4 vCPU, 16 GB of memory and a 64GB disk, and 1 executor per worker.

For the G.2X worker type, each worker provides 8 vCPU, 32 GB of memory and a 128GB disk, and 1 executor per worker.

getMLTransformResponse_inputRecordTables :: Lens' GetMLTransformResponse (Maybe [GlueTable])

A list of Glue table definitions used by the transform.

getMLTransformResponse_glueVersion :: Lens' GetMLTransformResponse (Maybe Text)

This value determines which version of Glue this machine learning transform is compatible with. Glue 1.0 is recommended for most customers. If the value is not set, the Glue compatibility defaults to Glue 0.9. For more information, see Glue Versions in the developer guide.

getMLTransformResponse_evaluationMetrics :: Lens' GetMLTransformResponse (Maybe EvaluationMetrics)

The latest evaluation metrics.

getMLTransformResponse_schema :: Lens' GetMLTransformResponse (Maybe [SchemaColumn])

The Map<Column, Type> object that represents the schema that this transform accepts. Has an upper bound of 100 columns.

getMLTransformResponse_role :: Lens' GetMLTransformResponse (Maybe Text)

The name or Amazon Resource Name (ARN) of the IAM role with the required permissions.

getMLTransformResponse_name :: Lens' GetMLTransformResponse (Maybe Text)

The unique name given to the transform when it was created.

getMLTransformResponse_parameters :: Lens' GetMLTransformResponse (Maybe TransformParameters)

The configuration parameters that are specific to the algorithm used.

getMLTransformResponse_maxRetries :: Lens' GetMLTransformResponse (Maybe Int)

The maximum number of times to retry a task for this transform after a task run fails.

getMLTransformResponse_maxCapacity :: Lens' GetMLTransformResponse (Maybe Double)

The number of Glue data processing units (DPUs) that are allocated to task runs for this transform. You can allocate from 2 to 100 DPUs; the default is 10. A DPU is a relative measure of processing power that consists of 4 vCPUs of compute capacity and 16 GB of memory. For more information, see the Glue pricing page.

When the WorkerType field is set to a value other than Standard, the MaxCapacity field is set automatically and becomes read-only.

getMLTransformResponse_timeout :: Lens' GetMLTransformResponse (Maybe Natural)

The timeout for a task run for this transform in minutes. This is the maximum time that a task run for this transform can consume resources before it is terminated and enters TIMEOUT status. The default is 2,880 minutes (48 hours).

getMLTransformResponse_transformEncryption :: Lens' GetMLTransformResponse (Maybe TransformEncryption)

The encryption-at-rest settings of the transform that apply to accessing user data. Machine learning transforms can access user data encrypted in Amazon S3 using KMS.

getMLTransformResponse_description :: Lens' GetMLTransformResponse (Maybe Text)

A description of the transform.

getMLTransformResponse_createdOn :: Lens' GetMLTransformResponse (Maybe UTCTime)

The date and time when the transform was created.

getMLTransformResponse_transformId :: Lens' GetMLTransformResponse (Maybe Text)

The unique identifier of the transform, generated at the time that the transform was created.

getMLTransformResponse_httpStatus :: Lens' GetMLTransformResponse Int

The response's http status code.