Amazonka.Glue.GetResourcePolicies

Description

Retrieves the resource policies set on individual resources by Resource Access Manager during cross-account permission grants. Also retrieves the Data Catalog resource policy.

If you enabled metadata encryption in Data Catalog settings, and you do not have permission on the KMS key, the operation can't return the Data Catalog resource policy.

This operation returns paginated results.

Creating a Request

data GetResourcePolicies

See: newGetResourcePolicies smart constructor.

Constructors

GetResourcePolicies'

Fields

nextToken :: Maybe Text

A continuation token, if this is a continuation request.

maxResults :: Maybe Natural

The maximum size of a list to return.

newGetResourcePolicies :: GetResourcePolicies

Create a value of GetResourcePolicies with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:nextToken:GetResourcePolicies', getResourcePolicies_nextToken - A continuation token, if this is a continuation request.

$sel:maxResults:GetResourcePolicies', getResourcePolicies_maxResults - The maximum size of a list to return.

Request Lenses

getResourcePolicies_nextToken :: Lens' GetResourcePolicies (Maybe Text)

A continuation token, if this is a continuation request.

getResourcePolicies_maxResults :: Lens' GetResourcePolicies (Maybe Natural)

The maximum size of a list to return.

Destructuring the Response

data GetResourcePoliciesResponse

See: newGetResourcePoliciesResponse smart constructor.

Constructors

GetResourcePoliciesResponse'

Fields

getResourcePoliciesResponseList :: Maybe [GluePolicy]

A list of the individual resource policies and the account-level resource policy.

nextToken :: Maybe Text

A continuation token, if the returned list does not contain the last resource policy available.

httpStatus :: Int

The response's http status code.

newGetResourcePoliciesResponse

Arguments

:: Int
$sel:httpStatus:GetResourcePoliciesResponse'

-> GetResourcePoliciesResponse

Create a value of GetResourcePoliciesResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:getResourcePoliciesResponseList:GetResourcePoliciesResponse', getResourcePoliciesResponse_getResourcePoliciesResponseList - A list of the individual resource policies and the account-level resource policy.

$sel:nextToken:GetResourcePolicies', getResourcePoliciesResponse_nextToken - A continuation token, if the returned list does not contain the last resource policy available.

$sel:httpStatus:GetResourcePoliciesResponse', getResourcePoliciesResponse_httpStatus - The response's http status code.

Response Lenses

getResourcePoliciesResponse_getResourcePoliciesResponseList :: Lens' GetResourcePoliciesResponse (Maybe [GluePolicy])

A list of the individual resource policies and the account-level resource policy.

getResourcePoliciesResponse_nextToken :: Lens' GetResourcePoliciesResponse (Maybe Text)

A continuation token, if the returned list does not contain the last resource policy available.

getResourcePoliciesResponse_httpStatus :: Lens' GetResourcePoliciesResponse Int

The response's http status code.