Amazonka.Glue.GetUserDefinedFunctions

Description

Retrieves multiple function definitions from the Data Catalog.

This operation returns paginated results.

Creating a Request

data GetUserDefinedFunctions

See: newGetUserDefinedFunctions smart constructor.

Constructors

GetUserDefinedFunctions'

Fields

catalogId :: Maybe Text

The ID of the Data Catalog where the functions to be retrieved are located. If none is provided, the Amazon Web Services account ID is used by default.

nextToken :: Maybe Text

A continuation token, if this is a continuation call.

databaseName :: Maybe Text

The name of the catalog database where the functions are located. If none is provided, functions from all the databases across the catalog will be returned.

maxResults :: Maybe Natural

The maximum number of functions to return in one response.

pattern' :: Text

An optional function-name pattern string that filters the function definitions returned.

newGetUserDefinedFunctions

Arguments

:: Text
$sel:pattern':GetUserDefinedFunctions'

-> GetUserDefinedFunctions

Create a value of GetUserDefinedFunctions with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:catalogId:GetUserDefinedFunctions', getUserDefinedFunctions_catalogId - The ID of the Data Catalog where the functions to be retrieved are located. If none is provided, the Amazon Web Services account ID is used by default.

$sel:nextToken:GetUserDefinedFunctions', getUserDefinedFunctions_nextToken - A continuation token, if this is a continuation call.

$sel:databaseName:GetUserDefinedFunctions', getUserDefinedFunctions_databaseName - The name of the catalog database where the functions are located. If none is provided, functions from all the databases across the catalog will be returned.

$sel:maxResults:GetUserDefinedFunctions', getUserDefinedFunctions_maxResults - The maximum number of functions to return in one response.

$sel:pattern':GetUserDefinedFunctions', getUserDefinedFunctions_pattern - An optional function-name pattern string that filters the function definitions returned.

Request Lenses

getUserDefinedFunctions_catalogId :: Lens' GetUserDefinedFunctions (Maybe Text)

The ID of the Data Catalog where the functions to be retrieved are located. If none is provided, the Amazon Web Services account ID is used by default.

getUserDefinedFunctions_nextToken :: Lens' GetUserDefinedFunctions (Maybe Text)

A continuation token, if this is a continuation call.

getUserDefinedFunctions_databaseName :: Lens' GetUserDefinedFunctions (Maybe Text)

The name of the catalog database where the functions are located. If none is provided, functions from all the databases across the catalog will be returned.

getUserDefinedFunctions_maxResults :: Lens' GetUserDefinedFunctions (Maybe Natural)

The maximum number of functions to return in one response.

getUserDefinedFunctions_pattern :: Lens' GetUserDefinedFunctions Text

An optional function-name pattern string that filters the function definitions returned.

Destructuring the Response

data GetUserDefinedFunctionsResponse

See: newGetUserDefinedFunctionsResponse smart constructor.

Constructors

GetUserDefinedFunctionsResponse'

Fields

nextToken :: Maybe Text

A continuation token, if the list of functions returned does not include the last requested function.

userDefinedFunctions :: Maybe [UserDefinedFunction]

A list of requested function definitions.

httpStatus :: Int

The response's http status code.

newGetUserDefinedFunctionsResponse

Arguments

:: Int
$sel:httpStatus:GetUserDefinedFunctionsResponse'

-> GetUserDefinedFunctionsResponse

Create a value of GetUserDefinedFunctionsResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:nextToken:GetUserDefinedFunctions', getUserDefinedFunctionsResponse_nextToken - A continuation token, if the list of functions returned does not include the last requested function.

$sel:userDefinedFunctions:GetUserDefinedFunctionsResponse', getUserDefinedFunctionsResponse_userDefinedFunctions - A list of requested function definitions.

$sel:httpStatus:GetUserDefinedFunctionsResponse', getUserDefinedFunctionsResponse_httpStatus - The response's http status code.

Response Lenses

getUserDefinedFunctionsResponse_nextToken :: Lens' GetUserDefinedFunctionsResponse (Maybe Text)

A continuation token, if the list of functions returned does not include the last requested function.

getUserDefinedFunctionsResponse_userDefinedFunctions :: Lens' GetUserDefinedFunctionsResponse (Maybe [UserDefinedFunction])

A list of requested function definitions.

getUserDefinedFunctionsResponse_httpStatus :: Lens' GetUserDefinedFunctionsResponse Int

The response's http status code.