Amazonka.Glue.ResetJobBookmark

Description

Resets a bookmark entry.

Creating a Request

data ResetJobBookmark

See: newResetJobBookmark smart constructor.

Constructors

ResetJobBookmark'

Fields

runId :: Maybe Text

The unique run identifier associated with this job run.

jobName :: Text

The name of the job in question.

newResetJobBookmark

Arguments

:: Text
$sel:jobName:ResetJobBookmark'

-> ResetJobBookmark

Create a value of ResetJobBookmark with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:runId:ResetJobBookmark', resetJobBookmark_runId - The unique run identifier associated with this job run.

$sel:jobName:ResetJobBookmark', resetJobBookmark_jobName - The name of the job in question.

Request Lenses

resetJobBookmark_runId :: Lens' ResetJobBookmark (Maybe Text)

The unique run identifier associated with this job run.

resetJobBookmark_jobName :: Lens' ResetJobBookmark Text

The name of the job in question.

Destructuring the Response

data ResetJobBookmarkResponse

See: newResetJobBookmarkResponse smart constructor.

Constructors

ResetJobBookmarkResponse'

Fields

jobBookmarkEntry :: Maybe JobBookmarkEntry

The reset bookmark entry.

httpStatus :: Int

The response's http status code.

newResetJobBookmarkResponse

Arguments

:: Int
$sel:httpStatus:ResetJobBookmarkResponse'

-> ResetJobBookmarkResponse

Create a value of ResetJobBookmarkResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:jobBookmarkEntry:ResetJobBookmarkResponse', resetJobBookmarkResponse_jobBookmarkEntry - The reset bookmark entry.

$sel:httpStatus:ResetJobBookmarkResponse', resetJobBookmarkResponse_httpStatus - The response's http status code.

Response Lenses

resetJobBookmarkResponse_jobBookmarkEntry :: Lens' ResetJobBookmarkResponse (Maybe JobBookmarkEntry)

The reset bookmark entry.

resetJobBookmarkResponse_httpStatus :: Lens' ResetJobBookmarkResponse Int

The response's http status code.