Amazonka.Glue.StartBlueprintRun

Description

Starts a new run of the specified blueprint.

Creating a Request

data StartBlueprintRun

See: newStartBlueprintRun smart constructor.

Constructors

StartBlueprintRun'

Fields

parameters :: Maybe Text

Specifies the parameters as a BlueprintParameters object.

blueprintName :: Text

The name of the blueprint.

roleArn :: Text

Specifies the IAM role used to create the workflow.

newStartBlueprintRun

Arguments

:: Text
$sel:blueprintName:StartBlueprintRun'

-> Text
$sel:roleArn:StartBlueprintRun'

-> StartBlueprintRun

Create a value of StartBlueprintRun with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:parameters:StartBlueprintRun', startBlueprintRun_parameters - Specifies the parameters as a BlueprintParameters object.

$sel:blueprintName:StartBlueprintRun', startBlueprintRun_blueprintName - The name of the blueprint.

$sel:roleArn:StartBlueprintRun', startBlueprintRun_roleArn - Specifies the IAM role used to create the workflow.

Request Lenses

startBlueprintRun_parameters :: Lens' StartBlueprintRun (Maybe Text)

Specifies the parameters as a BlueprintParameters object.

startBlueprintRun_blueprintName :: Lens' StartBlueprintRun Text

The name of the blueprint.

startBlueprintRun_roleArn :: Lens' StartBlueprintRun Text

Specifies the IAM role used to create the workflow.

Destructuring the Response

data StartBlueprintRunResponse

See: newStartBlueprintRunResponse smart constructor.

Constructors

StartBlueprintRunResponse'

Fields

runId :: Maybe Text

The run ID for this blueprint run.

httpStatus :: Int

The response's http status code.

newStartBlueprintRunResponse

Arguments

:: Int
$sel:httpStatus:StartBlueprintRunResponse'

-> StartBlueprintRunResponse

Create a value of StartBlueprintRunResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:runId:StartBlueprintRunResponse', startBlueprintRunResponse_runId - The run ID for this blueprint run.

$sel:httpStatus:StartBlueprintRunResponse', startBlueprintRunResponse_httpStatus - The response's http status code.

Response Lenses

startBlueprintRunResponse_runId :: Lens' StartBlueprintRunResponse (Maybe Text)

The run ID for this blueprint run.

startBlueprintRunResponse_httpStatus :: Lens' StartBlueprintRunResponse Int

The response's http status code.