Amazonka.Glue.Types.Blueprint

Description

Documentation

data Blueprint

The details of a blueprint.

See: newBlueprint smart constructor.

Constructors

Blueprint'

Fields

status :: Maybe BlueprintStatus

The status of the blueprint registration.

Creating — The blueprint registration is in progress.

Active — The blueprint has been successfully registered.

Updating — An update to the blueprint registration is in progress.

Failed — The blueprint registration failed.

parameterSpec :: Maybe Text

A JSON string that indicates the list of parameter specifications for the blueprint.

blueprintLocation :: Maybe Text

Specifies the path in Amazon S3 where the blueprint is published.

lastModifiedOn :: Maybe POSIX

The date and time the blueprint was last modified.

lastActiveDefinition :: Maybe LastActiveDefinition

When there are multiple versions of a blueprint and the latest version has some errors, this attribute indicates the last successful blueprint definition that is available with the service.

name :: Maybe Text

The name of the blueprint.

blueprintServiceLocation :: Maybe Text

Specifies a path in Amazon S3 where the blueprint is copied when you call CreateBlueprint/UpdateBlueprint to register the blueprint in Glue.

errorMessage :: Maybe Text

An error message.

description :: Maybe Text

The description of the blueprint.

createdOn :: Maybe POSIX

The date and time the blueprint was registered.

newBlueprint :: Blueprint

Create a value of Blueprint with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:status:Blueprint', blueprint_status - The status of the blueprint registration.

Creating — The blueprint registration is in progress.

Active — The blueprint has been successfully registered.

Updating — An update to the blueprint registration is in progress.

Failed — The blueprint registration failed.

$sel:parameterSpec:Blueprint', blueprint_parameterSpec - A JSON string that indicates the list of parameter specifications for the blueprint.

$sel:blueprintLocation:Blueprint', blueprint_blueprintLocation - Specifies the path in Amazon S3 where the blueprint is published.

$sel:lastModifiedOn:Blueprint', blueprint_lastModifiedOn - The date and time the blueprint was last modified.

$sel:lastActiveDefinition:Blueprint', blueprint_lastActiveDefinition - When there are multiple versions of a blueprint and the latest version has some errors, this attribute indicates the last successful blueprint definition that is available with the service.

$sel:name:Blueprint', blueprint_name - The name of the blueprint.

$sel:blueprintServiceLocation:Blueprint', blueprint_blueprintServiceLocation - Specifies a path in Amazon S3 where the blueprint is copied when you call CreateBlueprint/UpdateBlueprint to register the blueprint in Glue.

$sel:errorMessage:Blueprint', blueprint_errorMessage - An error message.

$sel:description:Blueprint', blueprint_description - The description of the blueprint.

$sel:createdOn:Blueprint', blueprint_createdOn - The date and time the blueprint was registered.

blueprint_status :: Lens' Blueprint (Maybe BlueprintStatus)

The status of the blueprint registration.

Creating — The blueprint registration is in progress.

Active — The blueprint has been successfully registered.

Updating — An update to the blueprint registration is in progress.

Failed — The blueprint registration failed.

blueprint_parameterSpec :: Lens' Blueprint (Maybe Text)

A JSON string that indicates the list of parameter specifications for the blueprint.

blueprint_blueprintLocation :: Lens' Blueprint (Maybe Text)

Specifies the path in Amazon S3 where the blueprint is published.

blueprint_lastModifiedOn :: Lens' Blueprint (Maybe UTCTime)

The date and time the blueprint was last modified.

blueprint_lastActiveDefinition :: Lens' Blueprint (Maybe LastActiveDefinition)

When there are multiple versions of a blueprint and the latest version has some errors, this attribute indicates the last successful blueprint definition that is available with the service.

blueprint_name :: Lens' Blueprint (Maybe Text)

The name of the blueprint.

blueprint_blueprintServiceLocation :: Lens' Blueprint (Maybe Text)

Specifies a path in Amazon S3 where the blueprint is copied when you call CreateBlueprint/UpdateBlueprint to register the blueprint in Glue.

blueprint_errorMessage :: Lens' Blueprint (Maybe Text)

An error message.

blueprint_description :: Lens' Blueprint (Maybe Text)

The description of the blueprint.

blueprint_createdOn :: Lens' Blueprint (Maybe UTCTime)

The date and time the blueprint was registered.