Amazonka.Glue.Types.CodeGenNodeArg

Description

Documentation

data CodeGenNodeArg

An argument or property of a node.

See: newCodeGenNodeArg smart constructor.

Constructors

CodeGenNodeArg'

Fields

param :: Maybe Bool

True if the value is used as a parameter.

name :: Text

The name of the argument or property.

value :: Text

The value of the argument or property.

newCodeGenNodeArg

Arguments

:: Text
$sel:name:CodeGenNodeArg'

-> Text
$sel:value:CodeGenNodeArg'

-> CodeGenNodeArg

Create a value of CodeGenNodeArg with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:param:CodeGenNodeArg', codeGenNodeArg_param - True if the value is used as a parameter.

$sel:name:CodeGenNodeArg', codeGenNodeArg_name - The name of the argument or property.

$sel:value:CodeGenNodeArg', codeGenNodeArg_value - The value of the argument or property.

codeGenNodeArg_param :: Lens' CodeGenNodeArg (Maybe Bool)

True if the value is used as a parameter.

codeGenNodeArg_name :: Lens' CodeGenNodeArg Text

The name of the argument or property.

codeGenNodeArg_value :: Lens' CodeGenNodeArg Text

The value of the argument or property.