Amazonka.Glue.Types.JobUpdate

Description

Documentation

data JobUpdate

Specifies information used to update an existing job definition. The previous job definition is completely overwritten by this information.

See: newJobUpdate smart constructor.

Constructors

JobUpdate'

Fields

numberOfWorkers :: Maybe Int

The number of workers of a defined workerType that are allocated when a job runs.

The maximum number of workers you can define are 299 for G.1X, and 149 for G.2X.

command :: Maybe JobCommand

The JobCommand that runs this job (required).

notificationProperty :: Maybe NotificationProperty

Specifies the configuration properties of a job notification.

connections :: Maybe ConnectionsList

The connections used for this job.

workerType :: Maybe WorkerType

The type of predefined worker that is allocated when a job runs. Accepts a value of Standard, G.1X, or G.2X.

For the Standard worker type, each worker provides 4 vCPU, 16 GB of memory and a 50GB disk, and 2 executors per worker.

For the G.1X worker type, each worker maps to 1 DPU (4 vCPU, 16 GB of memory, 64 GB disk), and provides 1 executor per worker. We recommend this worker type for memory-intensive jobs.

For the G.2X worker type, each worker maps to 2 DPU (8 vCPU, 32 GB of memory, 128 GB disk), and provides 1 executor per worker. We recommend this worker type for memory-intensive jobs.

securityConfiguration :: Maybe Text

The name of the SecurityConfiguration structure to be used with this job.

glueVersion :: Maybe Text

Glue version determines the versions of Apache Spark and Python that Glue supports. The Python version indicates the version supported for jobs of type Spark.

For more information about the available Glue versions and corresponding Spark and Python versions, see Glue version in the developer guide.

nonOverridableArguments :: Maybe (HashMap Text Text)

Non-overridable arguments for this job, specified as name-value pairs.

role' :: Maybe Text

The name or Amazon Resource Name (ARN) of the IAM role associated with this job (required).

logUri :: Maybe Text

This field is reserved for future use.

maxRetries :: Maybe Int

The maximum number of times to retry this job if it fails.

executionProperty :: Maybe ExecutionProperty

An ExecutionProperty specifying the maximum number of concurrent runs allowed for this job.

allocatedCapacity :: Maybe Int

This field is deprecated. Use MaxCapacity instead.

The number of Glue data processing units (DPUs) to allocate to this job. You can allocate from 2 to 100 DPUs; the default is 10. A DPU is a relative measure of processing power that consists of 4 vCPUs of compute capacity and 16 GB of memory. For more information, see the Glue pricing page.

maxCapacity :: Maybe Double

For Glue version 1.0 or earlier jobs, using the standard worker type, the number of Glue data processing units (DPUs) that can be allocated when this job runs. A DPU is a relative measure of processing power that consists of 4 vCPUs of compute capacity and 16 GB of memory. For more information, see the Glue pricing page.

Do not set Max Capacity if using WorkerType and NumberOfWorkers.

The value that can be allocated for MaxCapacity depends on whether you are running a Python shell job or an Apache Spark ETL job:

When you specify a Python shell job (JobCommand.Name="pythonshell"), you can allocate either 0.0625 or 1 DPU. The default is 0.0625 DPU.

When you specify an Apache Spark ETL job (JobCommand.Name="glueetl") or Apache Spark streaming ETL job (JobCommand.Name="gluestreaming"), you can allocate from 2 to 100 DPUs. The default is 10 DPUs. This job type cannot have a fractional DPU allocation.

For Glue version 2.0 jobs, you cannot instead specify a Maximum capacity. Instead, you should specify a Worker type and the Number of workers.

timeout :: Maybe Natural

The job timeout in minutes. This is the maximum time that a job run can consume resources before it is terminated and enters TIMEOUT status. The default is 2,880 minutes (48 hours).

defaultArguments :: Maybe (HashMap Text Text)

The default arguments for this job.

You can specify arguments here that your own job-execution script consumes, as well as arguments that Glue itself consumes.

For information about how to specify and consume your own Job arguments, see the Calling Glue APIs in Python topic in the developer guide.

For information about the key-value pairs that Glue consumes to set up your job, see the Special Parameters Used by Glue topic in the developer guide.

description :: Maybe Text

Description of the job being defined.

newJobUpdate :: JobUpdate

Create a value of JobUpdate with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:numberOfWorkers:JobUpdate', jobUpdate_numberOfWorkers - The number of workers of a defined workerType that are allocated when a job runs.

The maximum number of workers you can define are 299 for G.1X, and 149 for G.2X.

$sel:command:JobUpdate', jobUpdate_command - The JobCommand that runs this job (required).

$sel:notificationProperty:JobUpdate', jobUpdate_notificationProperty - Specifies the configuration properties of a job notification.

$sel:connections:JobUpdate', jobUpdate_connections - The connections used for this job.

$sel:workerType:JobUpdate', jobUpdate_workerType - The type of predefined worker that is allocated when a job runs. Accepts a value of Standard, G.1X, or G.2X.

For the Standard worker type, each worker provides 4 vCPU, 16 GB of memory and a 50GB disk, and 2 executors per worker.

For the G.1X worker type, each worker maps to 1 DPU (4 vCPU, 16 GB of memory, 64 GB disk), and provides 1 executor per worker. We recommend this worker type for memory-intensive jobs.

For the G.2X worker type, each worker maps to 2 DPU (8 vCPU, 32 GB of memory, 128 GB disk), and provides 1 executor per worker. We recommend this worker type for memory-intensive jobs.

$sel:securityConfiguration:JobUpdate', jobUpdate_securityConfiguration - The name of the SecurityConfiguration structure to be used with this job.

$sel:glueVersion:JobUpdate', jobUpdate_glueVersion - Glue version determines the versions of Apache Spark and Python that Glue supports. The Python version indicates the version supported for jobs of type Spark.

For more information about the available Glue versions and corresponding Spark and Python versions, see Glue version in the developer guide.

$sel:nonOverridableArguments:JobUpdate', jobUpdate_nonOverridableArguments - Non-overridable arguments for this job, specified as name-value pairs.

$sel:role':JobUpdate', jobUpdate_role - The name or Amazon Resource Name (ARN) of the IAM role associated with this job (required).

$sel:logUri:JobUpdate', jobUpdate_logUri - This field is reserved for future use.

$sel:maxRetries:JobUpdate', jobUpdate_maxRetries - The maximum number of times to retry this job if it fails.

$sel:executionProperty:JobUpdate', jobUpdate_executionProperty - An ExecutionProperty specifying the maximum number of concurrent runs allowed for this job.

$sel:allocatedCapacity:JobUpdate', jobUpdate_allocatedCapacity - This field is deprecated. Use MaxCapacity instead.

The number of Glue data processing units (DPUs) to allocate to this job. You can allocate from 2 to 100 DPUs; the default is 10. A DPU is a relative measure of processing power that consists of 4 vCPUs of compute capacity and 16 GB of memory. For more information, see the Glue pricing page.

$sel:maxCapacity:JobUpdate', jobUpdate_maxCapacity - For Glue version 1.0 or earlier jobs, using the standard worker type, the number of Glue data processing units (DPUs) that can be allocated when this job runs. A DPU is a relative measure of processing power that consists of 4 vCPUs of compute capacity and 16 GB of memory. For more information, see the Glue pricing page.

Do not set Max Capacity if using WorkerType and NumberOfWorkers.

The value that can be allocated for MaxCapacity depends on whether you are running a Python shell job or an Apache Spark ETL job:

When you specify a Python shell job (JobCommand.Name="pythonshell"), you can allocate either 0.0625 or 1 DPU. The default is 0.0625 DPU.

When you specify an Apache Spark ETL job (JobCommand.Name="glueetl") or Apache Spark streaming ETL job (JobCommand.Name="gluestreaming"), you can allocate from 2 to 100 DPUs. The default is 10 DPUs. This job type cannot have a fractional DPU allocation.

For Glue version 2.0 jobs, you cannot instead specify a Maximum capacity. Instead, you should specify a Worker type and the Number of workers.

$sel:timeout:JobUpdate', jobUpdate_timeout - The job timeout in minutes. This is the maximum time that a job run can consume resources before it is terminated and enters TIMEOUT status. The default is 2,880 minutes (48 hours).

$sel:defaultArguments:JobUpdate', jobUpdate_defaultArguments - The default arguments for this job.

You can specify arguments here that your own job-execution script consumes, as well as arguments that Glue itself consumes.

For information about how to specify and consume your own Job arguments, see the Calling Glue APIs in Python topic in the developer guide.

For information about the key-value pairs that Glue consumes to set up your job, see the Special Parameters Used by Glue topic in the developer guide.

$sel:description:JobUpdate', jobUpdate_description - Description of the job being defined.

jobUpdate_numberOfWorkers :: Lens' JobUpdate (Maybe Int)

The number of workers of a defined workerType that are allocated when a job runs.

The maximum number of workers you can define are 299 for G.1X, and 149 for G.2X.

jobUpdate_command :: Lens' JobUpdate (Maybe JobCommand)

The JobCommand that runs this job (required).

jobUpdate_notificationProperty :: Lens' JobUpdate (Maybe NotificationProperty)

Specifies the configuration properties of a job notification.

jobUpdate_connections :: Lens' JobUpdate (Maybe ConnectionsList)

The connections used for this job.

jobUpdate_workerType :: Lens' JobUpdate (Maybe WorkerType)

The type of predefined worker that is allocated when a job runs. Accepts a value of Standard, G.1X, or G.2X.

For the Standard worker type, each worker provides 4 vCPU, 16 GB of memory and a 50GB disk, and 2 executors per worker.

For the G.1X worker type, each worker maps to 1 DPU (4 vCPU, 16 GB of memory, 64 GB disk), and provides 1 executor per worker. We recommend this worker type for memory-intensive jobs.

For the G.2X worker type, each worker maps to 2 DPU (8 vCPU, 32 GB of memory, 128 GB disk), and provides 1 executor per worker. We recommend this worker type for memory-intensive jobs.

jobUpdate_securityConfiguration :: Lens' JobUpdate (Maybe Text)

The name of the SecurityConfiguration structure to be used with this job.

jobUpdate_glueVersion :: Lens' JobUpdate (Maybe Text)

Glue version determines the versions of Apache Spark and Python that Glue supports. The Python version indicates the version supported for jobs of type Spark.

For more information about the available Glue versions and corresponding Spark and Python versions, see Glue version in the developer guide.

jobUpdate_nonOverridableArguments :: Lens' JobUpdate (Maybe (HashMap Text Text))

Non-overridable arguments for this job, specified as name-value pairs.

jobUpdate_role :: Lens' JobUpdate (Maybe Text)

The name or Amazon Resource Name (ARN) of the IAM role associated with this job (required).

jobUpdate_logUri :: Lens' JobUpdate (Maybe Text)

This field is reserved for future use.

jobUpdate_maxRetries :: Lens' JobUpdate (Maybe Int)

The maximum number of times to retry this job if it fails.

jobUpdate_executionProperty :: Lens' JobUpdate (Maybe ExecutionProperty)

An ExecutionProperty specifying the maximum number of concurrent runs allowed for this job.

jobUpdate_allocatedCapacity :: Lens' JobUpdate (Maybe Int)

This field is deprecated. Use MaxCapacity instead.

The number of Glue data processing units (DPUs) to allocate to this job. You can allocate from 2 to 100 DPUs; the default is 10. A DPU is a relative measure of processing power that consists of 4 vCPUs of compute capacity and 16 GB of memory. For more information, see the Glue pricing page.

jobUpdate_maxCapacity :: Lens' JobUpdate (Maybe Double)

For Glue version 1.0 or earlier jobs, using the standard worker type, the number of Glue data processing units (DPUs) that can be allocated when this job runs. A DPU is a relative measure of processing power that consists of 4 vCPUs of compute capacity and 16 GB of memory. For more information, see the Glue pricing page.

Do not set Max Capacity if using WorkerType and NumberOfWorkers.

The value that can be allocated for MaxCapacity depends on whether you are running a Python shell job or an Apache Spark ETL job:

When you specify a Python shell job (JobCommand.Name="pythonshell"), you can allocate either 0.0625 or 1 DPU. The default is 0.0625 DPU.

When you specify an Apache Spark ETL job (JobCommand.Name="glueetl") or Apache Spark streaming ETL job (JobCommand.Name="gluestreaming"), you can allocate from 2 to 100 DPUs. The default is 10 DPUs. This job type cannot have a fractional DPU allocation.

For Glue version 2.0 jobs, you cannot instead specify a Maximum capacity. Instead, you should specify a Worker type and the Number of workers.

jobUpdate_timeout :: Lens' JobUpdate (Maybe Natural)

The job timeout in minutes. This is the maximum time that a job run can consume resources before it is terminated and enters TIMEOUT status. The default is 2,880 minutes (48 hours).

jobUpdate_defaultArguments :: Lens' JobUpdate (Maybe (HashMap Text Text))

The default arguments for this job.

You can specify arguments here that your own job-execution script consumes, as well as arguments that Glue itself consumes.

For information about how to specify and consume your own Job arguments, see the Calling Glue APIs in Python topic in the developer guide.

For information about the key-value pairs that Glue consumes to set up your job, see the Special Parameters Used by Glue topic in the developer guide.

jobUpdate_description :: Lens' JobUpdate (Maybe Text)

Description of the job being defined.