Amazonka.Glue.Types.MetadataInfo

Description

Documentation

data MetadataInfo

A structure containing metadata information for a schema version.

See: newMetadataInfo smart constructor.

Constructors

MetadataInfo'

Fields

createdTime :: Maybe Text

The time at which the entry was created.

otherMetadataValueList :: Maybe [OtherMetadataValueListItem]

Other metadata belonging to the same metadata key.

metadataValue :: Maybe Text

The metadata key’s corresponding value.

newMetadataInfo :: MetadataInfo

Create a value of MetadataInfo with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:createdTime:MetadataInfo', metadataInfo_createdTime - The time at which the entry was created.

$sel:otherMetadataValueList:MetadataInfo', metadataInfo_otherMetadataValueList - Other metadata belonging to the same metadata key.

$sel:metadataValue:MetadataInfo', metadataInfo_metadataValue - The metadata key’s corresponding value.

metadataInfo_createdTime :: Lens' MetadataInfo (Maybe Text)

The time at which the entry was created.

metadataInfo_otherMetadataValueList :: Lens' MetadataInfo (Maybe [OtherMetadataValueListItem])

Other metadata belonging to the same metadata key.

metadataInfo_metadataValue :: Lens' MetadataInfo (Maybe Text)

The metadata key’s corresponding value.