Amazonka.Glue.Types.Partition

Description

Documentation

data Partition

Represents a slice of table data.

See: newPartition smart constructor.

Constructors

Partition'

Fields

creationTime :: Maybe POSIX

The time at which the partition was created.

values :: Maybe [Text]

The values of the partition.

catalogId :: Maybe Text

The ID of the Data Catalog in which the partition resides.

lastAnalyzedTime :: Maybe POSIX

The last time at which column statistics were computed for this partition.

storageDescriptor :: Maybe StorageDescriptor

Provides information about the physical location where the partition is stored.

databaseName :: Maybe Text

The name of the catalog database in which to create the partition.

parameters :: Maybe (HashMap Text Text)

These key-value pairs define partition parameters.

lastAccessTime :: Maybe POSIX

The last time at which the partition was accessed.

tableName :: Maybe Text

The name of the database table in which to create the partition.

newPartition :: Partition

Create a value of Partition with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:creationTime:Partition', partition_creationTime - The time at which the partition was created.

$sel:values:Partition', partition_values - The values of the partition.

$sel:catalogId:Partition', partition_catalogId - The ID of the Data Catalog in which the partition resides.

$sel:lastAnalyzedTime:Partition', partition_lastAnalyzedTime - The last time at which column statistics were computed for this partition.

$sel:storageDescriptor:Partition', partition_storageDescriptor - Provides information about the physical location where the partition is stored.

$sel:databaseName:Partition', partition_databaseName - The name of the catalog database in which to create the partition.

$sel:parameters:Partition', partition_parameters - These key-value pairs define partition parameters.

$sel:lastAccessTime:Partition', partition_lastAccessTime - The last time at which the partition was accessed.

$sel:tableName:Partition', partition_tableName - The name of the database table in which to create the partition.

partition_creationTime :: Lens' Partition (Maybe UTCTime)

The time at which the partition was created.

partition_values :: Lens' Partition (Maybe [Text])

The values of the partition.

partition_catalogId :: Lens' Partition (Maybe Text)

The ID of the Data Catalog in which the partition resides.

partition_lastAnalyzedTime :: Lens' Partition (Maybe UTCTime)

The last time at which column statistics were computed for this partition.

partition_storageDescriptor :: Lens' Partition (Maybe StorageDescriptor)

Provides information about the physical location where the partition is stored.

partition_databaseName :: Lens' Partition (Maybe Text)

The name of the catalog database in which to create the partition.

partition_parameters :: Lens' Partition (Maybe (HashMap Text Text))

These key-value pairs define partition parameters.

partition_lastAccessTime :: Lens' Partition (Maybe UTCTime)

The last time at which the partition was accessed.

partition_tableName :: Lens' Partition (Maybe Text)

The name of the database table in which to create the partition.