Amazonka.Glue.Types.PartitionIndexDescriptor

Description

Documentation

data PartitionIndexDescriptor

A descriptor for a partition index in a table.

See: newPartitionIndexDescriptor smart constructor.

Constructors

PartitionIndexDescriptor'

Fields

backfillErrors :: Maybe [BackfillError]

A list of errors that can occur when registering partition indexes for an existing table.

indexName :: Text

The name of the partition index.

keys :: NonEmpty KeySchemaElement

A list of one or more keys, as KeySchemaElement structures, for the partition index.

indexStatus :: PartitionIndexStatus

The status of the partition index.

The possible statuses are:

CREATING: The index is being created. When an index is in a CREATING state, the index or its table cannot be deleted.

ACTIVE: The index creation succeeds.

FAILED: The index creation fails.

DELETING: The index is deleted from the list of indexes.

newPartitionIndexDescriptor

Arguments

:: Text
$sel:indexName:PartitionIndexDescriptor'

-> NonEmpty KeySchemaElement
$sel:keys:PartitionIndexDescriptor'

-> PartitionIndexStatus
$sel:indexStatus:PartitionIndexDescriptor'

-> PartitionIndexDescriptor

Create a value of PartitionIndexDescriptor with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:backfillErrors:PartitionIndexDescriptor', partitionIndexDescriptor_backfillErrors - A list of errors that can occur when registering partition indexes for an existing table.

$sel:indexName:PartitionIndexDescriptor', partitionIndexDescriptor_indexName - The name of the partition index.

$sel:keys:PartitionIndexDescriptor', partitionIndexDescriptor_keys - A list of one or more keys, as KeySchemaElement structures, for the partition index.

$sel:indexStatus:PartitionIndexDescriptor', partitionIndexDescriptor_indexStatus - The status of the partition index.

The possible statuses are:

CREATING: The index is being created. When an index is in a CREATING state, the index or its table cannot be deleted.

ACTIVE: The index creation succeeds.

FAILED: The index creation fails.

DELETING: The index is deleted from the list of indexes.

partitionIndexDescriptor_backfillErrors :: Lens' PartitionIndexDescriptor (Maybe [BackfillError])

A list of errors that can occur when registering partition indexes for an existing table.

partitionIndexDescriptor_indexName :: Lens' PartitionIndexDescriptor Text

The name of the partition index.

partitionIndexDescriptor_keys :: Lens' PartitionIndexDescriptor (NonEmpty KeySchemaElement)

A list of one or more keys, as KeySchemaElement structures, for the partition index.

partitionIndexDescriptor_indexStatus :: Lens' PartitionIndexDescriptor PartitionIndexStatus

The status of the partition index.

The possible statuses are:

CREATING: The index is being created. When an index is in a CREATING state, the index or its table cannot be deleted.

ACTIVE: The index creation succeeds.

FAILED: The index creation fails.

DELETING: The index is deleted from the list of indexes.