Amazonka.Glue.Types.TableInput

Description

Documentation

data TableInput

A structure used to define a table.

See: newTableInput smart constructor.

Constructors

TableInput'

Fields

retention :: Maybe Natural

The retention time for this table.

targetTable :: Maybe TableIdentifier

A TableIdentifier structure that describes a target table for resource linking.

tableType :: Maybe Text

The type of this table (EXTERNAL_TABLE, VIRTUAL_VIEW, etc.).

owner :: Maybe Text

The table owner.

viewOriginalText :: Maybe Text

If the table is a view, the original text of the view; otherwise null.

viewExpandedText :: Maybe Text

If the table is a view, the expanded text of the view; otherwise null.

lastAnalyzedTime :: Maybe POSIX

The last time that column statistics were computed for this table.

storageDescriptor :: Maybe StorageDescriptor

A storage descriptor containing information about the physical storage of this table.

parameters :: Maybe (HashMap Text Text)

These key-value pairs define properties associated with the table.

lastAccessTime :: Maybe POSIX

The last time that the table was accessed.

description :: Maybe Text

A description of the table.

partitionKeys :: Maybe [Column]

A list of columns by which the table is partitioned. Only primitive types are supported as partition keys.

When you create a table used by Amazon Athena, and you do not specify any partitionKeys, you must at least set the value of partitionKeys to an empty list. For example:

"PartitionKeys": []

name :: Text

The table name. For Hive compatibility, this is folded to lowercase when it is stored.

newTableInput

Arguments

:: Text
$sel:name:TableInput'

-> TableInput

Create a value of TableInput with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:retention:TableInput', tableInput_retention - The retention time for this table.

$sel:targetTable:TableInput', tableInput_targetTable - A TableIdentifier structure that describes a target table for resource linking.

$sel:tableType:TableInput', tableInput_tableType - The type of this table (EXTERNAL_TABLE, VIRTUAL_VIEW, etc.).

$sel:owner:TableInput', tableInput_owner - The table owner.

$sel:viewOriginalText:TableInput', tableInput_viewOriginalText - If the table is a view, the original text of the view; otherwise null.

$sel:viewExpandedText:TableInput', tableInput_viewExpandedText - If the table is a view, the expanded text of the view; otherwise null.

$sel:lastAnalyzedTime:TableInput', tableInput_lastAnalyzedTime - The last time that column statistics were computed for this table.

$sel:storageDescriptor:TableInput', tableInput_storageDescriptor - A storage descriptor containing information about the physical storage of this table.

$sel:parameters:TableInput', tableInput_parameters - These key-value pairs define properties associated with the table.

$sel:lastAccessTime:TableInput', tableInput_lastAccessTime - The last time that the table was accessed.

$sel:description:TableInput', tableInput_description - A description of the table.

$sel:partitionKeys:TableInput', tableInput_partitionKeys - A list of columns by which the table is partitioned. Only primitive types are supported as partition keys.

When you create a table used by Amazon Athena, and you do not specify any partitionKeys, you must at least set the value of partitionKeys to an empty list. For example:

"PartitionKeys": []

$sel:name:TableInput', tableInput_name - The table name. For Hive compatibility, this is folded to lowercase when it is stored.

tableInput_retention :: Lens' TableInput (Maybe Natural)

The retention time for this table.

tableInput_targetTable :: Lens' TableInput (Maybe TableIdentifier)

A TableIdentifier structure that describes a target table for resource linking.

tableInput_tableType :: Lens' TableInput (Maybe Text)

The type of this table (EXTERNAL_TABLE, VIRTUAL_VIEW, etc.).

tableInput_owner :: Lens' TableInput (Maybe Text)

The table owner.

tableInput_viewOriginalText :: Lens' TableInput (Maybe Text)

If the table is a view, the original text of the view; otherwise null.

tableInput_viewExpandedText :: Lens' TableInput (Maybe Text)

If the table is a view, the expanded text of the view; otherwise null.

tableInput_lastAnalyzedTime :: Lens' TableInput (Maybe UTCTime)

The last time that column statistics were computed for this table.

tableInput_storageDescriptor :: Lens' TableInput (Maybe StorageDescriptor)

A storage descriptor containing information about the physical storage of this table.

tableInput_parameters :: Lens' TableInput (Maybe (HashMap Text Text))

These key-value pairs define properties associated with the table.

tableInput_lastAccessTime :: Lens' TableInput (Maybe UTCTime)

The last time that the table was accessed.

tableInput_description :: Lens' TableInput (Maybe Text)

A description of the table.

tableInput_partitionKeys :: Lens' TableInput (Maybe [Column])

A list of columns by which the table is partitioned. Only primitive types are supported as partition keys.

When you create a table used by Amazon Athena, and you do not specify any partitionKeys, you must at least set the value of partitionKeys to an empty list. For example:

"PartitionKeys": []

tableInput_name :: Lens' TableInput Text

The table name. For Hive compatibility, this is folded to lowercase when it is stored.