Amazonka.Glue.Types.TransformFilterCriteria

Description

Documentation

data TransformFilterCriteria

The criteria used to filter the machine learning transforms.

See: newTransformFilterCriteria smart constructor.

Constructors

TransformFilterCriteria'

Fields

createdAfter :: Maybe POSIX

The time and date after which the transforms were created.

status :: Maybe TransformStatusType

Filters the list of machine learning transforms by the last known status of the transforms (to indicate whether a transform can be used or not). One of "NOT_READY", "READY", or "DELETING".

lastModifiedAfter :: Maybe POSIX

Filter on transforms last modified after this date.

lastModifiedBefore :: Maybe POSIX

Filter on transforms last modified before this date.

glueVersion :: Maybe Text

This value determines which version of Glue this machine learning transform is compatible with. Glue 1.0 is recommended for most customers. If the value is not set, the Glue compatibility defaults to Glue 0.9. For more information, see Glue Versions in the developer guide.

schema :: Maybe [SchemaColumn]

Filters on datasets with a specific schema. The Map<Column, Type> object is an array of key-value pairs representing the schema this transform accepts, where Column is the name of a column, and Type is the type of the data such as an integer or string. Has an upper bound of 100 columns.

transformType :: Maybe TransformType

The type of machine learning transform that is used to filter the machine learning transforms.

name :: Maybe Text

A unique transform name that is used to filter the machine learning transforms.

createdBefore :: Maybe POSIX

The time and date before which the transforms were created.

newTransformFilterCriteria :: TransformFilterCriteria

Create a value of TransformFilterCriteria with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:createdAfter:TransformFilterCriteria', transformFilterCriteria_createdAfter - The time and date after which the transforms were created.

$sel:status:TransformFilterCriteria', transformFilterCriteria_status - Filters the list of machine learning transforms by the last known status of the transforms (to indicate whether a transform can be used or not). One of "NOT_READY", "READY", or "DELETING".

$sel:lastModifiedAfter:TransformFilterCriteria', transformFilterCriteria_lastModifiedAfter - Filter on transforms last modified after this date.

$sel:lastModifiedBefore:TransformFilterCriteria', transformFilterCriteria_lastModifiedBefore - Filter on transforms last modified before this date.

$sel:glueVersion:TransformFilterCriteria', transformFilterCriteria_glueVersion - This value determines which version of Glue this machine learning transform is compatible with. Glue 1.0 is recommended for most customers. If the value is not set, the Glue compatibility defaults to Glue 0.9. For more information, see Glue Versions in the developer guide.

$sel:schema:TransformFilterCriteria', transformFilterCriteria_schema - Filters on datasets with a specific schema. The Map<Column, Type> object is an array of key-value pairs representing the schema this transform accepts, where Column is the name of a column, and Type is the type of the data such as an integer or string. Has an upper bound of 100 columns.

$sel:transformType:TransformFilterCriteria', transformFilterCriteria_transformType - The type of machine learning transform that is used to filter the machine learning transforms.

$sel:name:TransformFilterCriteria', transformFilterCriteria_name - A unique transform name that is used to filter the machine learning transforms.

$sel:createdBefore:TransformFilterCriteria', transformFilterCriteria_createdBefore - The time and date before which the transforms were created.

transformFilterCriteria_createdAfter :: Lens' TransformFilterCriteria (Maybe UTCTime)

The time and date after which the transforms were created.

transformFilterCriteria_status :: Lens' TransformFilterCriteria (Maybe TransformStatusType)

Filters the list of machine learning transforms by the last known status of the transforms (to indicate whether a transform can be used or not). One of "NOT_READY", "READY", or "DELETING".

transformFilterCriteria_lastModifiedAfter :: Lens' TransformFilterCriteria (Maybe UTCTime)

Filter on transforms last modified after this date.

transformFilterCriteria_lastModifiedBefore :: Lens' TransformFilterCriteria (Maybe UTCTime)

Filter on transforms last modified before this date.

transformFilterCriteria_glueVersion :: Lens' TransformFilterCriteria (Maybe Text)

This value determines which version of Glue this machine learning transform is compatible with. Glue 1.0 is recommended for most customers. If the value is not set, the Glue compatibility defaults to Glue 0.9. For more information, see Glue Versions in the developer guide.

transformFilterCriteria_schema :: Lens' TransformFilterCriteria (Maybe [SchemaColumn])

Filters on datasets with a specific schema. The Map<Column, Type> object is an array of key-value pairs representing the schema this transform accepts, where Column is the name of a column, and Type is the type of the data such as an integer or string. Has an upper bound of 100 columns.

transformFilterCriteria_transformType :: Lens' TransformFilterCriteria (Maybe TransformType)

The type of machine learning transform that is used to filter the machine learning transforms.

transformFilterCriteria_name :: Lens' TransformFilterCriteria (Maybe Text)

A unique transform name that is used to filter the machine learning transforms.

transformFilterCriteria_createdBefore :: Lens' TransformFilterCriteria (Maybe UTCTime)

The time and date before which the transforms were created.