Amazonka.Glue.Types.TriggerUpdate

Description

Documentation

data TriggerUpdate

A structure used to provide information used to update a trigger. This object updates the previous trigger definition by overwriting it completely.

See: newTriggerUpdate smart constructor.

Constructors

TriggerUpdate'

Fields

actions :: Maybe [Action]

The actions initiated by this trigger.

schedule :: Maybe Text

A cron expression used to specify the schedule (see Time-Based Schedules for Jobs and Crawlers. For example, to run something every day at 12:15 UTC, you would specify: cron(15 12 * * ? *).

predicate :: Maybe Predicate

The predicate of this trigger, which defines when it will fire.

name :: Maybe Text

Reserved for future use.

eventBatchingCondition :: Maybe EventBatchingCondition

Batch condition that must be met (specified number of events received or batch time window expired) before EventBridge event trigger fires.

description :: Maybe Text

A description of this trigger.

newTriggerUpdate :: TriggerUpdate

Create a value of TriggerUpdate with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:actions:TriggerUpdate', triggerUpdate_actions - The actions initiated by this trigger.

$sel:schedule:TriggerUpdate', triggerUpdate_schedule - A cron expression used to specify the schedule (see Time-Based Schedules for Jobs and Crawlers. For example, to run something every day at 12:15 UTC, you would specify: cron(15 12 * * ? *).

$sel:predicate:TriggerUpdate', triggerUpdate_predicate - The predicate of this trigger, which defines when it will fire.

$sel:name:TriggerUpdate', triggerUpdate_name - Reserved for future use.

$sel:eventBatchingCondition:TriggerUpdate', triggerUpdate_eventBatchingCondition - Batch condition that must be met (specified number of events received or batch time window expired) before EventBridge event trigger fires.

$sel:description:TriggerUpdate', triggerUpdate_description - A description of this trigger.

triggerUpdate_actions :: Lens' TriggerUpdate (Maybe [Action])

The actions initiated by this trigger.

triggerUpdate_schedule :: Lens' TriggerUpdate (Maybe Text)

A cron expression used to specify the schedule (see Time-Based Schedules for Jobs and Crawlers. For example, to run something every day at 12:15 UTC, you would specify: cron(15 12 * * ? *).

triggerUpdate_predicate :: Lens' TriggerUpdate (Maybe Predicate)

The predicate of this trigger, which defines when it will fire.

triggerUpdate_name :: Lens' TriggerUpdate (Maybe Text)

Reserved for future use.

triggerUpdate_eventBatchingCondition :: Lens' TriggerUpdate (Maybe EventBatchingCondition)

Batch condition that must be met (specified number of events received or batch time window expired) before EventBridge event trigger fires.

triggerUpdate_description :: Lens' TriggerUpdate (Maybe Text)

A description of this trigger.