Amazonka.Glue.Types.UpdateGrokClassifierRequest

Description

Documentation

data UpdateGrokClassifierRequest

Specifies a grok classifier to update when passed to UpdateClassifier.

See: newUpdateGrokClassifierRequest smart constructor.

Constructors

UpdateGrokClassifierRequest'

Fields

classification :: Maybe Text

An identifier of the data format that the classifier matches, such as Twitter, JSON, Omniture logs, Amazon CloudWatch Logs, and so on.

customPatterns :: Maybe Text

Optional custom grok patterns used by this classifier.

grokPattern :: Maybe Text

The grok pattern used by this classifier.

name :: Text

The name of the GrokClassifier.

newUpdateGrokClassifierRequest

Arguments

:: Text
$sel:name:UpdateGrokClassifierRequest'

-> UpdateGrokClassifierRequest

Create a value of UpdateGrokClassifierRequest with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:classification:UpdateGrokClassifierRequest', updateGrokClassifierRequest_classification - An identifier of the data format that the classifier matches, such as Twitter, JSON, Omniture logs, Amazon CloudWatch Logs, and so on.

$sel:customPatterns:UpdateGrokClassifierRequest', updateGrokClassifierRequest_customPatterns - Optional custom grok patterns used by this classifier.

$sel:grokPattern:UpdateGrokClassifierRequest', updateGrokClassifierRequest_grokPattern - The grok pattern used by this classifier.

$sel:name:UpdateGrokClassifierRequest', updateGrokClassifierRequest_name - The name of the GrokClassifier.

updateGrokClassifierRequest_classification :: Lens' UpdateGrokClassifierRequest (Maybe Text)

An identifier of the data format that the classifier matches, such as Twitter, JSON, Omniture logs, Amazon CloudWatch Logs, and so on.

updateGrokClassifierRequest_customPatterns :: Lens' UpdateGrokClassifierRequest (Maybe Text)

Optional custom grok patterns used by this classifier.

updateGrokClassifierRequest_grokPattern :: Lens' UpdateGrokClassifierRequest (Maybe Text)

The grok pattern used by this classifier.

updateGrokClassifierRequest_name :: Lens' UpdateGrokClassifierRequest Text

The name of the GrokClassifier.