Amazonka.Glue.UpdateColumnStatisticsForTable

Description

Creates or updates table statistics of columns.

The Identity and Access Management (IAM) permission required for this operation is UpdateTable.

Creating a Request

data UpdateColumnStatisticsForTable

See: newUpdateColumnStatisticsForTable smart constructor.

Constructors

UpdateColumnStatisticsForTable'

Fields

catalogId :: Maybe Text

The ID of the Data Catalog where the partitions in question reside. If none is supplied, the Amazon Web Services account ID is used by default.

databaseName :: Text

The name of the catalog database where the partitions reside.

tableName :: Text

The name of the partitions' table.

columnStatisticsList :: [ColumnStatistics]

A list of the column statistics.

newUpdateColumnStatisticsForTable

Arguments

:: Text
$sel:databaseName:UpdateColumnStatisticsForTable'

-> Text
$sel:tableName:UpdateColumnStatisticsForTable'

-> UpdateColumnStatisticsForTable

Create a value of UpdateColumnStatisticsForTable with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:catalogId:UpdateColumnStatisticsForTable', updateColumnStatisticsForTable_catalogId - The ID of the Data Catalog where the partitions in question reside. If none is supplied, the Amazon Web Services account ID is used by default.

$sel:databaseName:UpdateColumnStatisticsForTable', updateColumnStatisticsForTable_databaseName - The name of the catalog database where the partitions reside.

$sel:tableName:UpdateColumnStatisticsForTable', updateColumnStatisticsForTable_tableName - The name of the partitions' table.

$sel:columnStatisticsList:UpdateColumnStatisticsForTable', updateColumnStatisticsForTable_columnStatisticsList - A list of the column statistics.

Request Lenses

updateColumnStatisticsForTable_catalogId :: Lens' UpdateColumnStatisticsForTable (Maybe Text)

The ID of the Data Catalog where the partitions in question reside. If none is supplied, the Amazon Web Services account ID is used by default.

updateColumnStatisticsForTable_databaseName :: Lens' UpdateColumnStatisticsForTable Text

The name of the catalog database where the partitions reside.

updateColumnStatisticsForTable_tableName :: Lens' UpdateColumnStatisticsForTable Text

The name of the partitions' table.

updateColumnStatisticsForTable_columnStatisticsList :: Lens' UpdateColumnStatisticsForTable [ColumnStatistics]

A list of the column statistics.

Destructuring the Response

data UpdateColumnStatisticsForTableResponse

See: newUpdateColumnStatisticsForTableResponse smart constructor.

Constructors

UpdateColumnStatisticsForTableResponse'

Fields

errors :: Maybe [ColumnStatisticsError]

List of ColumnStatisticsErrors.

httpStatus :: Int

The response's http status code.

newUpdateColumnStatisticsForTableResponse

Arguments

:: Int
$sel:httpStatus:UpdateColumnStatisticsForTableResponse'

-> UpdateColumnStatisticsForTableResponse

Create a value of UpdateColumnStatisticsForTableResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:errors:UpdateColumnStatisticsForTableResponse', updateColumnStatisticsForTableResponse_errors - List of ColumnStatisticsErrors.

$sel:httpStatus:UpdateColumnStatisticsForTableResponse', updateColumnStatisticsForTableResponse_httpStatus - The response's http status code.

Response Lenses

updateColumnStatisticsForTableResponse_errors :: Lens' UpdateColumnStatisticsForTableResponse (Maybe [ColumnStatisticsError])

List of ColumnStatisticsErrors.

updateColumnStatisticsForTableResponse_httpStatus :: Lens' UpdateColumnStatisticsForTableResponse Int

The response's http status code.