Amazonka.Glue.UpdateUserDefinedFunction

Description

Updates an existing function definition in the Data Catalog.

Creating a Request

data UpdateUserDefinedFunction

See: newUpdateUserDefinedFunction smart constructor.

Constructors

UpdateUserDefinedFunction'

Fields

catalogId :: Maybe Text

The ID of the Data Catalog where the function to be updated is located. If none is provided, the Amazon Web Services account ID is used by default.

databaseName :: Text

The name of the catalog database where the function to be updated is located.

functionName :: Text

The name of the function.

functionInput :: UserDefinedFunctionInput

A FunctionInput object that redefines the function in the Data Catalog.

newUpdateUserDefinedFunction

Arguments

:: Text
$sel:databaseName:UpdateUserDefinedFunction'

-> Text
$sel:functionName:UpdateUserDefinedFunction'

-> UserDefinedFunctionInput
$sel:functionInput:UpdateUserDefinedFunction'

-> UpdateUserDefinedFunction

Create a value of UpdateUserDefinedFunction with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:catalogId:UpdateUserDefinedFunction', updateUserDefinedFunction_catalogId - The ID of the Data Catalog where the function to be updated is located. If none is provided, the Amazon Web Services account ID is used by default.

$sel:databaseName:UpdateUserDefinedFunction', updateUserDefinedFunction_databaseName - The name of the catalog database where the function to be updated is located.

$sel:functionName:UpdateUserDefinedFunction', updateUserDefinedFunction_functionName - The name of the function.

$sel:functionInput:UpdateUserDefinedFunction', updateUserDefinedFunction_functionInput - A FunctionInput object that redefines the function in the Data Catalog.

Request Lenses

updateUserDefinedFunction_catalogId :: Lens' UpdateUserDefinedFunction (Maybe Text)

The ID of the Data Catalog where the function to be updated is located. If none is provided, the Amazon Web Services account ID is used by default.

updateUserDefinedFunction_databaseName :: Lens' UpdateUserDefinedFunction Text

The name of the catalog database where the function to be updated is located.

updateUserDefinedFunction_functionName :: Lens' UpdateUserDefinedFunction Text

The name of the function.

updateUserDefinedFunction_functionInput :: Lens' UpdateUserDefinedFunction UserDefinedFunctionInput

A FunctionInput object that redefines the function in the Data Catalog.

Destructuring the Response

data UpdateUserDefinedFunctionResponse

See: newUpdateUserDefinedFunctionResponse smart constructor.

Constructors

UpdateUserDefinedFunctionResponse'

Fields

httpStatus :: Int

The response's http status code.

newUpdateUserDefinedFunctionResponse

Arguments

:: Int
$sel:httpStatus:UpdateUserDefinedFunctionResponse'

-> UpdateUserDefinedFunctionResponse

Create a value of UpdateUserDefinedFunctionResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:httpStatus:UpdateUserDefinedFunctionResponse', updateUserDefinedFunctionResponse_httpStatus - The response's http status code.

Response Lenses

updateUserDefinedFunctionResponse_httpStatus :: Lens' UpdateUserDefinedFunctionResponse Int

The response's http status code.