Amazonka.Grafana.CreateWorkspace

Description

Creates a workspace. In a workspace, you can create Grafana dashboards and visualizations to analyze your metrics, logs, and traces. You don't have to build, package, or deploy any hardware to run the Grafana server.

Don't use CreateWorkspace to modify an existing workspace. Instead, use UpdateWorkspace.

Creating a Request

data CreateWorkspace

See: newCreateWorkspace smart constructor.

Constructors

CreateWorkspace'

Fields

workspaceRoleArn :: Maybe (Sensitive Text)

The workspace needs an IAM role that grants permissions to the Amazon Web Services resources that the workspace will view data from. If you already have a role that you want to use, specify it here. If you omit this field and you specify some Amazon Web Services resources in workspaceDataSources or workspaceNotificationDestinations, a new IAM role with the necessary permissions is automatically created.

clientToken :: Maybe Text

A unique, case-sensitive, user-provided identifier to ensure the idempotency of the request.

workspaceDataSources :: Maybe [DataSourceType]

Specify the Amazon Web Services data sources that you want to be queried in this workspace. Specifying these data sources here enables Amazon Managed Grafana to create IAM roles and permissions that allow Amazon Managed Grafana to read data from these sources. You must still add them as data sources in the Grafana console in the workspace.

If you don't specify a data source here, you can still add it as a data source in the workspace console later. However, you will then have to manually configure permissions for it.

workspaceNotificationDestinations :: Maybe [NotificationDestinationType]

Specify the Amazon Web Services notification channels that you plan to use in this workspace. Specifying these data sources here enables Amazon Managed Grafana to create IAM roles and permissions that allow Amazon Managed Grafana to use these channels.

workspaceName :: Maybe (Sensitive Text)

The name for the workspace. It does not have to be unique.

organizationRoleName :: Maybe (Sensitive Text)

The name of an IAM role that already exists to use with Organizations to access Amazon Web Services data sources and notification channels in other accounts in an organization.

workspaceDescription :: Maybe (Sensitive Text)

A description for the workspace. This is used only to help you identify this workspace.

stackSetName :: Maybe Text

The name of the CloudFormation stack set to use to generate IAM roles to be used for this workspace.

workspaceOrganizationalUnits :: Maybe (Sensitive [Text])

Specifies the organizational units that this workspace is allowed to use data sources from, if this workspace is in an account that is part of an organization.

accountAccessType :: AccountAccessType

Specifies whether the workspace can access Amazon Web Services resources in this Amazon Web Services account only, or whether it can also access Amazon Web Services resources in other accounts in the same organization. If you specify ORGANIZATION, you must specify which organizational units the workspace can access in the workspaceOrganizationalUnits parameter.

authenticationProviders :: [AuthenticationProviderTypes]

Specifies whether this workspace uses SAML 2.0, Amazon Web Services Single Sign On, or both to authenticate users for using the Grafana console within a workspace. For more information, see User authentication in Amazon Managed Grafana.

permissionType :: PermissionType

If you specify Service Managed, Amazon Managed Grafana automatically creates the IAM roles and provisions the permissions that the workspace needs to use Amazon Web Services data sources and notification channels.

If you specify CUSTOMER_MANAGED, you will manage those roles and permissions yourself. If you are creating this workspace in a member account of an organization that is not a delegated administrator account, and you want the workspace to access data sources in other Amazon Web Services accounts in the organization, you must choose CUSTOMER_MANAGED.

For more information, see Amazon Managed Grafana permissions and policies for Amazon Web Services data sources and notification channels

newCreateWorkspace

Arguments

:: AccountAccessType
$sel:accountAccessType:CreateWorkspace'

-> PermissionType
$sel:permissionType:CreateWorkspace'

-> CreateWorkspace

Create a value of CreateWorkspace with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:workspaceRoleArn:CreateWorkspace', createWorkspace_workspaceRoleArn - The workspace needs an IAM role that grants permissions to the Amazon Web Services resources that the workspace will view data from. If you already have a role that you want to use, specify it here. If you omit this field and you specify some Amazon Web Services resources in workspaceDataSources or workspaceNotificationDestinations, a new IAM role with the necessary permissions is automatically created.

$sel:clientToken:CreateWorkspace', createWorkspace_clientToken - A unique, case-sensitive, user-provided identifier to ensure the idempotency of the request.

$sel:workspaceDataSources:CreateWorkspace', createWorkspace_workspaceDataSources - Specify the Amazon Web Services data sources that you want to be queried in this workspace. Specifying these data sources here enables Amazon Managed Grafana to create IAM roles and permissions that allow Amazon Managed Grafana to read data from these sources. You must still add them as data sources in the Grafana console in the workspace.

If you don't specify a data source here, you can still add it as a data source in the workspace console later. However, you will then have to manually configure permissions for it.

$sel:workspaceNotificationDestinations:CreateWorkspace', createWorkspace_workspaceNotificationDestinations - Specify the Amazon Web Services notification channels that you plan to use in this workspace. Specifying these data sources here enables Amazon Managed Grafana to create IAM roles and permissions that allow Amazon Managed Grafana to use these channels.

$sel:workspaceName:CreateWorkspace', createWorkspace_workspaceName - The name for the workspace. It does not have to be unique.

$sel:organizationRoleName:CreateWorkspace', createWorkspace_organizationRoleName - The name of an IAM role that already exists to use with Organizations to access Amazon Web Services data sources and notification channels in other accounts in an organization.

$sel:workspaceDescription:CreateWorkspace', createWorkspace_workspaceDescription - A description for the workspace. This is used only to help you identify this workspace.

$sel:stackSetName:CreateWorkspace', createWorkspace_stackSetName - The name of the CloudFormation stack set to use to generate IAM roles to be used for this workspace.

$sel:workspaceOrganizationalUnits:CreateWorkspace', createWorkspace_workspaceOrganizationalUnits - Specifies the organizational units that this workspace is allowed to use data sources from, if this workspace is in an account that is part of an organization.

$sel:accountAccessType:CreateWorkspace', createWorkspace_accountAccessType - Specifies whether the workspace can access Amazon Web Services resources in this Amazon Web Services account only, or whether it can also access Amazon Web Services resources in other accounts in the same organization. If you specify ORGANIZATION, you must specify which organizational units the workspace can access in the workspaceOrganizationalUnits parameter.

$sel:authenticationProviders:CreateWorkspace', createWorkspace_authenticationProviders - Specifies whether this workspace uses SAML 2.0, Amazon Web Services Single Sign On, or both to authenticate users for using the Grafana console within a workspace. For more information, see User authentication in Amazon Managed Grafana.

$sel:permissionType:CreateWorkspace', createWorkspace_permissionType - If you specify Service Managed, Amazon Managed Grafana automatically creates the IAM roles and provisions the permissions that the workspace needs to use Amazon Web Services data sources and notification channels.

If you specify CUSTOMER_MANAGED, you will manage those roles and permissions yourself. If you are creating this workspace in a member account of an organization that is not a delegated administrator account, and you want the workspace to access data sources in other Amazon Web Services accounts in the organization, you must choose CUSTOMER_MANAGED.

For more information, see Amazon Managed Grafana permissions and policies for Amazon Web Services data sources and notification channels

Request Lenses

createWorkspace_workspaceRoleArn :: Lens' CreateWorkspace (Maybe Text)

The workspace needs an IAM role that grants permissions to the Amazon Web Services resources that the workspace will view data from. If you already have a role that you want to use, specify it here. If you omit this field and you specify some Amazon Web Services resources in workspaceDataSources or workspaceNotificationDestinations, a new IAM role with the necessary permissions is automatically created.

createWorkspace_clientToken :: Lens' CreateWorkspace (Maybe Text)

A unique, case-sensitive, user-provided identifier to ensure the idempotency of the request.

createWorkspace_workspaceDataSources :: Lens' CreateWorkspace (Maybe [DataSourceType])

Specify the Amazon Web Services data sources that you want to be queried in this workspace. Specifying these data sources here enables Amazon Managed Grafana to create IAM roles and permissions that allow Amazon Managed Grafana to read data from these sources. You must still add them as data sources in the Grafana console in the workspace.

If you don't specify a data source here, you can still add it as a data source in the workspace console later. However, you will then have to manually configure permissions for it.

createWorkspace_workspaceNotificationDestinations :: Lens' CreateWorkspace (Maybe [NotificationDestinationType])

Specify the Amazon Web Services notification channels that you plan to use in this workspace. Specifying these data sources here enables Amazon Managed Grafana to create IAM roles and permissions that allow Amazon Managed Grafana to use these channels.

createWorkspace_workspaceName :: Lens' CreateWorkspace (Maybe Text)

The name for the workspace. It does not have to be unique.

createWorkspace_organizationRoleName :: Lens' CreateWorkspace (Maybe Text)

The name of an IAM role that already exists to use with Organizations to access Amazon Web Services data sources and notification channels in other accounts in an organization.

createWorkspace_workspaceDescription :: Lens' CreateWorkspace (Maybe Text)

A description for the workspace. This is used only to help you identify this workspace.

createWorkspace_stackSetName :: Lens' CreateWorkspace (Maybe Text)

The name of the CloudFormation stack set to use to generate IAM roles to be used for this workspace.

createWorkspace_workspaceOrganizationalUnits :: Lens' CreateWorkspace (Maybe [Text])

Specifies the organizational units that this workspace is allowed to use data sources from, if this workspace is in an account that is part of an organization.

createWorkspace_accountAccessType :: Lens' CreateWorkspace AccountAccessType

Specifies whether the workspace can access Amazon Web Services resources in this Amazon Web Services account only, or whether it can also access Amazon Web Services resources in other accounts in the same organization. If you specify ORGANIZATION, you must specify which organizational units the workspace can access in the workspaceOrganizationalUnits parameter.

createWorkspace_authenticationProviders :: Lens' CreateWorkspace [AuthenticationProviderTypes]

Specifies whether this workspace uses SAML 2.0, Amazon Web Services Single Sign On, or both to authenticate users for using the Grafana console within a workspace. For more information, see User authentication in Amazon Managed Grafana.

createWorkspace_permissionType :: Lens' CreateWorkspace PermissionType

If you specify Service Managed, Amazon Managed Grafana automatically creates the IAM roles and provisions the permissions that the workspace needs to use Amazon Web Services data sources and notification channels.

If you specify CUSTOMER_MANAGED, you will manage those roles and permissions yourself. If you are creating this workspace in a member account of an organization that is not a delegated administrator account, and you want the workspace to access data sources in other Amazon Web Services accounts in the organization, you must choose CUSTOMER_MANAGED.

For more information, see Amazon Managed Grafana permissions and policies for Amazon Web Services data sources and notification channels

Destructuring the Response

data CreateWorkspaceResponse

See: newCreateWorkspaceResponse smart constructor.

Constructors

CreateWorkspaceResponse'

Fields

httpStatus :: Int

The response's http status code.

workspace :: WorkspaceDescription

A structure containing data about the workspace that was created.

newCreateWorkspaceResponse

Arguments

:: Int
$sel:httpStatus:CreateWorkspaceResponse'

-> WorkspaceDescription
$sel:workspace:CreateWorkspaceResponse'

-> CreateWorkspaceResponse

Create a value of CreateWorkspaceResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:httpStatus:CreateWorkspaceResponse', createWorkspaceResponse_httpStatus - The response's http status code.

$sel:workspace:CreateWorkspaceResponse', createWorkspaceResponse_workspace - A structure containing data about the workspace that was created.

Response Lenses

createWorkspaceResponse_httpStatus :: Lens' CreateWorkspaceResponse Int

The response's http status code.

createWorkspaceResponse_workspace :: Lens' CreateWorkspaceResponse WorkspaceDescription

A structure containing data about the workspace that was created.