Amazonka.Grafana.Types

Description

Service Configuration

defaultService :: Service

API version 2020-08-18 of the Amazon Managed Grafana SDK configuration.

Errors

_ValidationException :: AsError a => Getting (First ServiceError) a ServiceError

The value of a parameter in the request caused an error.

_AccessDeniedException :: AsError a => Getting (First ServiceError) a ServiceError

You do not have sufficient permissions to perform this action.

_ConflictException :: AsError a => Getting (First ServiceError) a ServiceError

A resource was in an inconsistent state during an update or a deletion.

_ServiceQuotaExceededException :: AsError a => Getting (First ServiceError) a ServiceError

The request would cause a service quota to be exceeded.

_ThrottlingException :: AsError a => Getting (First ServiceError) a ServiceError

The request was denied because of request throttling. Retry the request.

_InternalServerException :: AsError a => Getting (First ServiceError) a ServiceError

Unexpected error while processing the request. Retry the request.

_ResourceNotFoundException :: AsError a => Getting (First ServiceError) a ServiceError

The request references a resource that does not exist.

AccountAccessType

newtype AccountAccessType

Constructors

AccountAccessType'

Fields

fromAccountAccessType :: Text

Bundled Patterns

pattern AccountAccessType_CURRENT_ACCOUNT :: AccountAccessType

pattern AccountAccessType_ORGANIZATION :: AccountAccessType

AuthenticationProviderTypes

newtype AuthenticationProviderTypes

Constructors

AuthenticationProviderTypes'

Fields

fromAuthenticationProviderTypes :: Text

Bundled Patterns

pattern AuthenticationProviderTypes_AWS_SSO :: AuthenticationProviderTypes

pattern AuthenticationProviderTypes_SAML :: AuthenticationProviderTypes

DataSourceType

newtype DataSourceType

Constructors

DataSourceType'

Fields

fromDataSourceType :: Text

Bundled Patterns

pattern DataSourceType_AMAZON_OPENSEARCH_SERVICE :: DataSourceType

pattern DataSourceType_CLOUDWATCH :: DataSourceType

pattern DataSourceType_PROMETHEUS :: DataSourceType

pattern DataSourceType_SITEWISE :: DataSourceType

pattern DataSourceType_TIMESTREAM :: DataSourceType

pattern DataSourceType_XRAY :: DataSourceType

LicenseType

newtype LicenseType

Constructors

LicenseType'

Fields

fromLicenseType :: Text

Bundled Patterns

pattern LicenseType_ENTERPRISE :: LicenseType

pattern LicenseType_ENTERPRISE_FREE_TRIAL :: LicenseType

NotificationDestinationType

newtype NotificationDestinationType

Constructors

NotificationDestinationType'

Fields

fromNotificationDestinationType :: Text

Bundled Patterns

pattern NotificationDestinationType_SNS :: NotificationDestinationType

PermissionType

newtype PermissionType

Constructors

PermissionType'

Fields

fromPermissionType :: Text

Bundled Patterns

pattern PermissionType_CUSTOMER_MANAGED :: PermissionType

pattern PermissionType_SERVICE_MANAGED :: PermissionType

Role

newtype Role

Constructors

Role'

Fields

fromRole :: Text

Bundled Patterns

pattern Role_ADMIN :: Role

pattern Role_EDITOR :: Role

SamlConfigurationStatus

newtype SamlConfigurationStatus

Constructors

SamlConfigurationStatus'

Fields

fromSamlConfigurationStatus :: Text

Bundled Patterns

pattern SamlConfigurationStatus_CONFIGURED :: SamlConfigurationStatus

pattern SamlConfigurationStatus_NOT_CONFIGURED :: SamlConfigurationStatus

UpdateAction

newtype UpdateAction

Constructors

UpdateAction'

Fields

fromUpdateAction :: Text

Bundled Patterns

pattern UpdateAction_ADD :: UpdateAction

pattern UpdateAction_REVOKE :: UpdateAction

UserType

newtype UserType

Constructors

UserType'

Fields

fromUserType :: Text

Bundled Patterns

pattern UserType_SSO_GROUP :: UserType

pattern UserType_SSO_USER :: UserType

WorkspaceStatus

newtype WorkspaceStatus

Constructors

WorkspaceStatus'

Fields

fromWorkspaceStatus :: Text

Bundled Patterns

pattern WorkspaceStatus_ACTIVE :: WorkspaceStatus

pattern WorkspaceStatus_CREATING :: WorkspaceStatus

pattern WorkspaceStatus_CREATION_FAILED :: WorkspaceStatus

pattern WorkspaceStatus_DELETING :: WorkspaceStatus

pattern WorkspaceStatus_DELETION_FAILED :: WorkspaceStatus

pattern WorkspaceStatus_FAILED :: WorkspaceStatus

pattern WorkspaceStatus_LICENSE_REMOVAL_FAILED :: WorkspaceStatus

pattern WorkspaceStatus_UPDATE_FAILED :: WorkspaceStatus

pattern WorkspaceStatus_UPDATING :: WorkspaceStatus

pattern WorkspaceStatus_UPGRADE_FAILED :: WorkspaceStatus

pattern WorkspaceStatus_UPGRADING :: WorkspaceStatus

AssertionAttributes

data AssertionAttributes

A structure that defines which attributes in the IdP assertion are to be used to define information about the users authenticated by the IdP to use the workspace.

See: newAssertionAttributes smart constructor.

Constructors

AssertionAttributes'

Fields

email :: Maybe Text

The name of the attribute within the SAML assertion to use as the email names for SAML users.

groups :: Maybe Text

The name of the attribute within the SAML assertion to use as the user full "friendly" names for user groups.

org :: Maybe Text

The name of the attribute within the SAML assertion to use as the user full "friendly" names for the users' organizations.

role' :: Maybe Text

The name of the attribute within the SAML assertion to use as the user roles.

name :: Maybe Text

The name of the attribute within the SAML assertion to use as the user full "friendly" names for SAML users.

login :: Maybe Text

The name of the attribute within the SAML assertion to use as the login names for SAML users.

newAssertionAttributes :: AssertionAttributes

Create a value of AssertionAttributes with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:email:AssertionAttributes', assertionAttributes_email - The name of the attribute within the SAML assertion to use as the email names for SAML users.

$sel:groups:AssertionAttributes', assertionAttributes_groups - The name of the attribute within the SAML assertion to use as the user full "friendly" names for user groups.

$sel:org:AssertionAttributes', assertionAttributes_org - The name of the attribute within the SAML assertion to use as the user full "friendly" names for the users' organizations.

$sel:role':AssertionAttributes', assertionAttributes_role - The name of the attribute within the SAML assertion to use as the user roles.

$sel:name:AssertionAttributes', assertionAttributes_name - The name of the attribute within the SAML assertion to use as the user full "friendly" names for SAML users.

$sel:login:AssertionAttributes', assertionAttributes_login - The name of the attribute within the SAML assertion to use as the login names for SAML users.

assertionAttributes_email :: Lens' AssertionAttributes (Maybe Text)

The name of the attribute within the SAML assertion to use as the email names for SAML users.

assertionAttributes_groups :: Lens' AssertionAttributes (Maybe Text)

The name of the attribute within the SAML assertion to use as the user full "friendly" names for user groups.

assertionAttributes_org :: Lens' AssertionAttributes (Maybe Text)

The name of the attribute within the SAML assertion to use as the user full "friendly" names for the users' organizations.

assertionAttributes_role :: Lens' AssertionAttributes (Maybe Text)

The name of the attribute within the SAML assertion to use as the user roles.

assertionAttributes_name :: Lens' AssertionAttributes (Maybe Text)

The name of the attribute within the SAML assertion to use as the user full "friendly" names for SAML users.

assertionAttributes_login :: Lens' AssertionAttributes (Maybe Text)

The name of the attribute within the SAML assertion to use as the login names for SAML users.

AuthenticationDescription

data AuthenticationDescription

A structure containing information about the user authentication methods used by the workspace.

See: newAuthenticationDescription smart constructor.

Constructors

AuthenticationDescription'

Fields

awsSso :: Maybe AwsSsoAuthentication

A structure containing information about how this workspace works with Amazon Web Services SSO.

saml :: Maybe SamlAuthentication

A structure containing information about how this workspace works with SAML, including what attributes within the assertion are to be mapped to user information in the workspace.

providers :: [AuthenticationProviderTypes]

Specifies whether this workspace uses Amazon Web Services SSO, SAML, or both methods to authenticate users to use the Grafana console in the Amazon Managed Grafana workspace.

newAuthenticationDescription :: AuthenticationDescription

Create a value of AuthenticationDescription with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:awsSso:AuthenticationDescription', authenticationDescription_awsSso - A structure containing information about how this workspace works with Amazon Web Services SSO.

$sel:saml:AuthenticationDescription', authenticationDescription_saml - A structure containing information about how this workspace works with SAML, including what attributes within the assertion are to be mapped to user information in the workspace.

$sel:providers:AuthenticationDescription', authenticationDescription_providers - Specifies whether this workspace uses Amazon Web Services SSO, SAML, or both methods to authenticate users to use the Grafana console in the Amazon Managed Grafana workspace.

authenticationDescription_awsSso :: Lens' AuthenticationDescription (Maybe AwsSsoAuthentication)

A structure containing information about how this workspace works with Amazon Web Services SSO.

authenticationDescription_saml :: Lens' AuthenticationDescription (Maybe SamlAuthentication)

A structure containing information about how this workspace works with SAML, including what attributes within the assertion are to be mapped to user information in the workspace.

authenticationDescription_providers :: Lens' AuthenticationDescription [AuthenticationProviderTypes]

Specifies whether this workspace uses Amazon Web Services SSO, SAML, or both methods to authenticate users to use the Grafana console in the Amazon Managed Grafana workspace.

AuthenticationSummary

data AuthenticationSummary

A structure that describes whether the workspace uses SAML, Amazon Web Services SSO, or both methods for user authentication, and whether that authentication is fully configured.

See: newAuthenticationSummary smart constructor.

Constructors

AuthenticationSummary'

Fields

samlConfigurationStatus :: Maybe SamlConfigurationStatus

Specifies whether the workplace's user authentication method is fully configured.

providers :: [AuthenticationProviderTypes]

Specifies whether the workspace uses SAML, Amazon Web Services SSO, or both methods for user authentication.

newAuthenticationSummary :: AuthenticationSummary

Create a value of AuthenticationSummary with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:samlConfigurationStatus:AuthenticationSummary', authenticationSummary_samlConfigurationStatus - Specifies whether the workplace's user authentication method is fully configured.

$sel:providers:AuthenticationSummary', authenticationSummary_providers - Specifies whether the workspace uses SAML, Amazon Web Services SSO, or both methods for user authentication.

authenticationSummary_samlConfigurationStatus :: Lens' AuthenticationSummary (Maybe SamlConfigurationStatus)

Specifies whether the workplace's user authentication method is fully configured.

authenticationSummary_providers :: Lens' AuthenticationSummary [AuthenticationProviderTypes]

Specifies whether the workspace uses SAML, Amazon Web Services SSO, or both methods for user authentication.

AwsSsoAuthentication

data AwsSsoAuthentication

A structure containing information about how this workspace works with Amazon Web Services SSO.

See: newAwsSsoAuthentication smart constructor.

Constructors

AwsSsoAuthentication'

Fields

ssoClientId :: Maybe Text

The ID of the Amazon Web Services SSO-managed application that is created by Amazon Managed Grafana.

newAwsSsoAuthentication :: AwsSsoAuthentication

Create a value of AwsSsoAuthentication with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:ssoClientId:AwsSsoAuthentication', awsSsoAuthentication_ssoClientId - The ID of the Amazon Web Services SSO-managed application that is created by Amazon Managed Grafana.

awsSsoAuthentication_ssoClientId :: Lens' AwsSsoAuthentication (Maybe Text)

The ID of the Amazon Web Services SSO-managed application that is created by Amazon Managed Grafana.

IdpMetadata

data IdpMetadata

A structure containing the identity provider (IdP) metadata used to integrate the identity provider with this workspace. You can specify the metadata either by providing a URL to its location in the url parameter, or by specifying the full metadata in XML format in the xml parameter.

See: newIdpMetadata smart constructor.

Constructors

IdpMetadata'

Fields

url :: Maybe Text

The URL of the location containing the metadata.

xml :: Maybe Text

The actual full metadata file, in XML format.

newIdpMetadata :: IdpMetadata

Create a value of IdpMetadata with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:url:IdpMetadata', idpMetadata_url - The URL of the location containing the metadata.

$sel:xml:IdpMetadata', idpMetadata_xml - The actual full metadata file, in XML format.

idpMetadata_url :: Lens' IdpMetadata (Maybe Text)

The URL of the location containing the metadata.

idpMetadata_xml :: Lens' IdpMetadata (Maybe Text)

The actual full metadata file, in XML format.

PermissionEntry

data PermissionEntry

A structure containing the identity of one user or group and the Admin or Editor role that they have.

See: newPermissionEntry smart constructor.

Constructors

PermissionEntry'

Fields

role' :: Role

Specifies whether the user or group has the Admin or Editor role.

user :: User

A structure with the ID of the user or group with this role.

newPermissionEntry

Arguments

:: Role
$sel:role':PermissionEntry'

-> User
$sel:user:PermissionEntry'

-> PermissionEntry

Create a value of PermissionEntry with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:role':PermissionEntry', permissionEntry_role - Specifies whether the user or group has the Admin or Editor role.

$sel:user:PermissionEntry', permissionEntry_user - A structure with the ID of the user or group with this role.

permissionEntry_role :: Lens' PermissionEntry Role

Specifies whether the user or group has the Admin or Editor role.

permissionEntry_user :: Lens' PermissionEntry User

A structure with the ID of the user or group with this role.

RoleValues

data RoleValues

This structure defines which groups defined in the SAML assertion attribute are to be mapped to the Grafana Admin and Editor roles in the workspace.

See: newRoleValues smart constructor.

Constructors

RoleValues'

Fields

admin :: Maybe [Text]

A list of groups from the SAML assertion attribute to grant the Grafana Admin role to.

editor :: Maybe [Text]

A list of groups from the SAML assertion attribute to grant the Grafana Editor role to.

newRoleValues :: RoleValues

Create a value of RoleValues with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:admin:RoleValues', roleValues_admin - A list of groups from the SAML assertion attribute to grant the Grafana Admin role to.

$sel:editor:RoleValues', roleValues_editor - A list of groups from the SAML assertion attribute to grant the Grafana Editor role to.

roleValues_admin :: Lens' RoleValues (Maybe [Text])

A list of groups from the SAML assertion attribute to grant the Grafana Admin role to.

roleValues_editor :: Lens' RoleValues (Maybe [Text])

A list of groups from the SAML assertion attribute to grant the Grafana Editor role to.

SamlAuthentication

data SamlAuthentication

A structure containing information about how this workspace works with SAML.

See: newSamlAuthentication smart constructor.

Constructors

SamlAuthentication'

Fields

configuration :: Maybe SamlConfiguration

A structure containing details about how this workspace works with SAML.

status :: SamlConfigurationStatus

Specifies whether the workspace's SAML configuration is complete.

newSamlAuthentication

Arguments

:: SamlConfigurationStatus
$sel:status:SamlAuthentication'

-> SamlAuthentication

Create a value of SamlAuthentication with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:configuration:SamlAuthentication', samlAuthentication_configuration - A structure containing details about how this workspace works with SAML.

$sel:status:SamlAuthentication', samlAuthentication_status - Specifies whether the workspace's SAML configuration is complete.

samlAuthentication_configuration :: Lens' SamlAuthentication (Maybe SamlConfiguration)

A structure containing details about how this workspace works with SAML.

samlAuthentication_status :: Lens' SamlAuthentication SamlConfigurationStatus

Specifies whether the workspace's SAML configuration is complete.

SamlConfiguration

data SamlConfiguration

A structure containing information about how this workspace works with SAML.

See: newSamlConfiguration smart constructor.

Constructors

SamlConfiguration'

Fields

loginValidityDuration :: Maybe Int

How long a sign-on session by a SAML user is valid, before the user has to sign on again.

assertionAttributes :: Maybe AssertionAttributes

A structure that defines which attributes in the SAML assertion are to be used to define information about the users authenticated by that IdP to use the workspace.

allowedOrganizations :: Maybe [Text]

Lists which organizations defined in the SAML assertion are allowed to use the Amazon Managed Grafana workspace. If this is empty, all organizations in the assertion attribute have access.

roleValues :: Maybe RoleValues

A structure containing arrays that map group names in the SAML assertion to the Grafana Admin and Editor roles in the workspace.

idpMetadata :: IdpMetadata

A structure containing the identity provider (IdP) metadata used to integrate the identity provider with this workspace.

newSamlConfiguration

Arguments

:: IdpMetadata
$sel:idpMetadata:SamlConfiguration'

-> SamlConfiguration

Create a value of SamlConfiguration with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:loginValidityDuration:SamlConfiguration', samlConfiguration_loginValidityDuration - How long a sign-on session by a SAML user is valid, before the user has to sign on again.

$sel:assertionAttributes:SamlConfiguration', samlConfiguration_assertionAttributes - A structure that defines which attributes in the SAML assertion are to be used to define information about the users authenticated by that IdP to use the workspace.

$sel:allowedOrganizations:SamlConfiguration', samlConfiguration_allowedOrganizations - Lists which organizations defined in the SAML assertion are allowed to use the Amazon Managed Grafana workspace. If this is empty, all organizations in the assertion attribute have access.

$sel:roleValues:SamlConfiguration', samlConfiguration_roleValues - A structure containing arrays that map group names in the SAML assertion to the Grafana Admin and Editor roles in the workspace.

$sel:idpMetadata:SamlConfiguration', samlConfiguration_idpMetadata - A structure containing the identity provider (IdP) metadata used to integrate the identity provider with this workspace.

samlConfiguration_loginValidityDuration :: Lens' SamlConfiguration (Maybe Int)

How long a sign-on session by a SAML user is valid, before the user has to sign on again.

samlConfiguration_assertionAttributes :: Lens' SamlConfiguration (Maybe AssertionAttributes)

A structure that defines which attributes in the SAML assertion are to be used to define information about the users authenticated by that IdP to use the workspace.

samlConfiguration_allowedOrganizations :: Lens' SamlConfiguration (Maybe [Text])

Lists which organizations defined in the SAML assertion are allowed to use the Amazon Managed Grafana workspace. If this is empty, all organizations in the assertion attribute have access.

samlConfiguration_roleValues :: Lens' SamlConfiguration (Maybe RoleValues)

A structure containing arrays that map group names in the SAML assertion to the Grafana Admin and Editor roles in the workspace.

samlConfiguration_idpMetadata :: Lens' SamlConfiguration IdpMetadata

A structure containing the identity provider (IdP) metadata used to integrate the identity provider with this workspace.

UpdateError

data UpdateError

A structure containing information about one error encountered while performing an UpdatePermissions operation.

See: newUpdateError smart constructor.

Constructors

UpdateError'

Fields

causedBy :: UpdateInstruction

Specifies which permission update caused the error.

code :: Natural

The error code.

message :: Text

The message for this error.

newUpdateError

Arguments

:: UpdateInstruction
$sel:causedBy:UpdateError'

-> Natural
$sel:code:UpdateError'

-> Text
$sel:message:UpdateError'

-> UpdateError

Create a value of UpdateError with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:causedBy:UpdateError', updateError_causedBy - Specifies which permission update caused the error.

$sel:code:UpdateError', updateError_code - The error code.

$sel:message:UpdateError', updateError_message - The message for this error.

updateError_causedBy :: Lens' UpdateError UpdateInstruction

Specifies which permission update caused the error.

updateError_code :: Lens' UpdateError Natural

The error code.

updateError_message :: Lens' UpdateError Text

The message for this error.

UpdateInstruction

data UpdateInstruction

Contains the instructions for one Grafana role permission update in a UpdatePermissions operation.

See: newUpdateInstruction smart constructor.

Constructors

UpdateInstruction'

Fields

action :: UpdateAction

Specifies whether this update is to add or revoke role permissions.

role' :: Role

The role to add or revoke for the user or the group specified in users.

users :: [User]

A structure that specifies the user or group to add or revoke the role for.

newUpdateInstruction

Arguments

:: UpdateAction
$sel:action:UpdateInstruction'

-> Role
$sel:role':UpdateInstruction'

-> UpdateInstruction

Create a value of UpdateInstruction with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:action:UpdateInstruction', updateInstruction_action - Specifies whether this update is to add or revoke role permissions.

$sel:role':UpdateInstruction', updateInstruction_role - The role to add or revoke for the user or the group specified in users.

$sel:users:UpdateInstruction', updateInstruction_users - A structure that specifies the user or group to add or revoke the role for.

updateInstruction_action :: Lens' UpdateInstruction UpdateAction

Specifies whether this update is to add or revoke role permissions.

updateInstruction_role :: Lens' UpdateInstruction Role

The role to add or revoke for the user or the group specified in users.

updateInstruction_users :: Lens' UpdateInstruction [User]

A structure that specifies the user or group to add or revoke the role for.

User

data User

A structure that specifies one user or group in the workspace.

See: newUser smart constructor.

Constructors

User'

Fields

id :: Text

The ID of the user or group.

type' :: UserType

Specifies whether this is a single user or a group.

newUser

Arguments

:: Text
$sel:id:User'

-> UserType
$sel:type':User'

-> User

Create a value of User with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:id:User', user_id - The ID of the user or group.

$sel:type':User', user_type - Specifies whether this is a single user or a group.

user_id :: Lens' User Text

The ID of the user or group.

user_type :: Lens' User UserType

Specifies whether this is a single user or a group.

WorkspaceDescription

data WorkspaceDescription

A structure containing information about an Amazon Managed Grafana workspace in your account.

See: newWorkspaceDescription smart constructor.

Constructors

WorkspaceDescription'

Fields

workspaceRoleArn :: Maybe (Sensitive Text)

The IAM role that grants permissions to the Amazon Web Services resources that the workspace will view data from. This role must already exist.

freeTrialExpiration :: Maybe POSIX

If this workspace is currently in the free trial period for Grafana Enterprise, this value specifies when that free trial ends.

licenseType :: Maybe LicenseType

Specifies whether this workspace has a full Grafana Enterprise license or a free trial license.

permissionType :: Maybe PermissionType

If this is Service Managed, Amazon Managed Grafana automatically creates the IAM roles and provisions the permissions that the workspace needs to use Amazon Web Services data sources and notification channels.

If this is CUSTOMER_MANAGED, you manage those roles and permissions yourself. If you are creating this workspace in a member account of an organization and that account is not a delegated administrator account, and you want the workspace to access data sources in other Amazon Web Services accounts in the organization, you must choose CUSTOMER_MANAGED.

For more information, see Amazon Managed Grafana permissions and policies for Amazon Web Services data sources and notification channels

name :: Maybe (Sensitive Text)

The name of the workspace.

notificationDestinations :: Maybe [NotificationDestinationType]

The Amazon Web Services notification channels that Amazon Managed Grafana can automatically create IAM roles and permissions for, to allow Amazon Managed Grafana to use these channels.

accountAccessType :: Maybe AccountAccessType

Specifies whether the workspace can access Amazon Web Services resources in this Amazon Web Services account only, or whether it can also access Amazon Web Services resources in other accounts in the same organization. If this is ORGANIZATION, the workspaceOrganizationalUnits parameter specifies which organizational units the workspace can access.

licenseExpiration :: Maybe POSIX

If this workspace has a full Grafana Enterprise license, this specifies when the license ends and will need to be renewed.

organizationRoleName :: Maybe (Sensitive Text)

The name of the IAM role that is used to access resources through Organizations.

stackSetName :: Maybe Text

The name of the CloudFormation stack set that is used to generate IAM roles to be used for this workspace.

organizationalUnits :: Maybe (Sensitive [Text])

Specifies the organizational units that this workspace is allowed to use data sources from, if this workspace is in an account that is part of an organization.

description :: Maybe (Sensitive Text)

The user-defined description of the workspace.

freeTrialConsumed :: Maybe Bool

Specifies whether this workspace has already fully used its free trial for Grafana Enterprise.

authentication :: AuthenticationSummary

A structure that describes whether the workspace uses SAML, Amazon Web Services SSO, or both methods for user authentication.

created :: POSIX

The date that the workspace was created.

dataSources :: [DataSourceType]

Specifies the Amazon Web Services data sources that have been configured to have IAM roles and permissions created to allow Amazon Managed Grafana to read data from these sources.

endpoint :: Text

The URL that users can use to access the Grafana console in the workspace.

grafanaVersion :: Text

The version of Grafana supported in this workspace.

id :: Text

The unique ID of this workspace.

modified :: POSIX

The most recent date that the workspace was modified.

status :: WorkspaceStatus

The current status of the workspace.

newWorkspaceDescription

Arguments

:: AuthenticationSummary
$sel:authentication:WorkspaceDescription'

-> UTCTime
$sel:created:WorkspaceDescription'

-> Text
$sel:endpoint:WorkspaceDescription'

-> Text
$sel:grafanaVersion:WorkspaceDescription'

-> Text
$sel:id:WorkspaceDescription'

-> UTCTime
$sel:modified:WorkspaceDescription'

-> WorkspaceStatus
$sel:status:WorkspaceDescription'

-> WorkspaceDescription

Create a value of WorkspaceDescription with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:workspaceRoleArn:WorkspaceDescription', workspaceDescription_workspaceRoleArn - The IAM role that grants permissions to the Amazon Web Services resources that the workspace will view data from. This role must already exist.

$sel:freeTrialExpiration:WorkspaceDescription', workspaceDescription_freeTrialExpiration - If this workspace is currently in the free trial period for Grafana Enterprise, this value specifies when that free trial ends.

$sel:licenseType:WorkspaceDescription', workspaceDescription_licenseType - Specifies whether this workspace has a full Grafana Enterprise license or a free trial license.

$sel:permissionType:WorkspaceDescription', workspaceDescription_permissionType - If this is Service Managed, Amazon Managed Grafana automatically creates the IAM roles and provisions the permissions that the workspace needs to use Amazon Web Services data sources and notification channels.

If this is CUSTOMER_MANAGED, you manage those roles and permissions yourself. If you are creating this workspace in a member account of an organization and that account is not a delegated administrator account, and you want the workspace to access data sources in other Amazon Web Services accounts in the organization, you must choose CUSTOMER_MANAGED.

For more information, see Amazon Managed Grafana permissions and policies for Amazon Web Services data sources and notification channels

$sel:name:WorkspaceDescription', workspaceDescription_name - The name of the workspace.

$sel:notificationDestinations:WorkspaceDescription', workspaceDescription_notificationDestinations - The Amazon Web Services notification channels that Amazon Managed Grafana can automatically create IAM roles and permissions for, to allow Amazon Managed Grafana to use these channels.

$sel:accountAccessType:WorkspaceDescription', workspaceDescription_accountAccessType - Specifies whether the workspace can access Amazon Web Services resources in this Amazon Web Services account only, or whether it can also access Amazon Web Services resources in other accounts in the same organization. If this is ORGANIZATION, the workspaceOrganizationalUnits parameter specifies which organizational units the workspace can access.

$sel:licenseExpiration:WorkspaceDescription', workspaceDescription_licenseExpiration - If this workspace has a full Grafana Enterprise license, this specifies when the license ends and will need to be renewed.

$sel:organizationRoleName:WorkspaceDescription', workspaceDescription_organizationRoleName - The name of the IAM role that is used to access resources through Organizations.

$sel:stackSetName:WorkspaceDescription', workspaceDescription_stackSetName - The name of the CloudFormation stack set that is used to generate IAM roles to be used for this workspace.

$sel:organizationalUnits:WorkspaceDescription', workspaceDescription_organizationalUnits - Specifies the organizational units that this workspace is allowed to use data sources from, if this workspace is in an account that is part of an organization.

$sel:description:WorkspaceDescription', workspaceDescription_description - The user-defined description of the workspace.

$sel:freeTrialConsumed:WorkspaceDescription', workspaceDescription_freeTrialConsumed - Specifies whether this workspace has already fully used its free trial for Grafana Enterprise.

$sel:authentication:WorkspaceDescription', workspaceDescription_authentication - A structure that describes whether the workspace uses SAML, Amazon Web Services SSO, or both methods for user authentication.

$sel:created:WorkspaceDescription', workspaceDescription_created - The date that the workspace was created.

$sel:dataSources:WorkspaceDescription', workspaceDescription_dataSources - Specifies the Amazon Web Services data sources that have been configured to have IAM roles and permissions created to allow Amazon Managed Grafana to read data from these sources.

$sel:endpoint:WorkspaceDescription', workspaceDescription_endpoint - The URL that users can use to access the Grafana console in the workspace.

$sel:grafanaVersion:WorkspaceDescription', workspaceDescription_grafanaVersion - The version of Grafana supported in this workspace.

$sel:id:WorkspaceDescription', workspaceDescription_id - The unique ID of this workspace.

$sel:modified:WorkspaceDescription', workspaceDescription_modified - The most recent date that the workspace was modified.

$sel:status:WorkspaceDescription', workspaceDescription_status - The current status of the workspace.

workspaceDescription_workspaceRoleArn :: Lens' WorkspaceDescription (Maybe Text)

The IAM role that grants permissions to the Amazon Web Services resources that the workspace will view data from. This role must already exist.

workspaceDescription_freeTrialExpiration :: Lens' WorkspaceDescription (Maybe UTCTime)

If this workspace is currently in the free trial period for Grafana Enterprise, this value specifies when that free trial ends.

workspaceDescription_licenseType :: Lens' WorkspaceDescription (Maybe LicenseType)

Specifies whether this workspace has a full Grafana Enterprise license or a free trial license.

workspaceDescription_permissionType :: Lens' WorkspaceDescription (Maybe PermissionType)

If this is Service Managed, Amazon Managed Grafana automatically creates the IAM roles and provisions the permissions that the workspace needs to use Amazon Web Services data sources and notification channels.

If this is CUSTOMER_MANAGED, you manage those roles and permissions yourself. If you are creating this workspace in a member account of an organization and that account is not a delegated administrator account, and you want the workspace to access data sources in other Amazon Web Services accounts in the organization, you must choose CUSTOMER_MANAGED.

For more information, see Amazon Managed Grafana permissions and policies for Amazon Web Services data sources and notification channels

workspaceDescription_name :: Lens' WorkspaceDescription (Maybe Text)

The name of the workspace.

workspaceDescription_notificationDestinations :: Lens' WorkspaceDescription (Maybe [NotificationDestinationType])

The Amazon Web Services notification channels that Amazon Managed Grafana can automatically create IAM roles and permissions for, to allow Amazon Managed Grafana to use these channels.

workspaceDescription_accountAccessType :: Lens' WorkspaceDescription (Maybe AccountAccessType)

Specifies whether the workspace can access Amazon Web Services resources in this Amazon Web Services account only, or whether it can also access Amazon Web Services resources in other accounts in the same organization. If this is ORGANIZATION, the workspaceOrganizationalUnits parameter specifies which organizational units the workspace can access.

workspaceDescription_licenseExpiration :: Lens' WorkspaceDescription (Maybe UTCTime)

If this workspace has a full Grafana Enterprise license, this specifies when the license ends and will need to be renewed.

workspaceDescription_organizationRoleName :: Lens' WorkspaceDescription (Maybe Text)

The name of the IAM role that is used to access resources through Organizations.

workspaceDescription_stackSetName :: Lens' WorkspaceDescription (Maybe Text)

The name of the CloudFormation stack set that is used to generate IAM roles to be used for this workspace.

workspaceDescription_organizationalUnits :: Lens' WorkspaceDescription (Maybe [Text])

Specifies the organizational units that this workspace is allowed to use data sources from, if this workspace is in an account that is part of an organization.

workspaceDescription_description :: Lens' WorkspaceDescription (Maybe Text)

The user-defined description of the workspace.

workspaceDescription_freeTrialConsumed :: Lens' WorkspaceDescription (Maybe Bool)

Specifies whether this workspace has already fully used its free trial for Grafana Enterprise.

workspaceDescription_authentication :: Lens' WorkspaceDescription AuthenticationSummary

A structure that describes whether the workspace uses SAML, Amazon Web Services SSO, or both methods for user authentication.

workspaceDescription_created :: Lens' WorkspaceDescription UTCTime

The date that the workspace was created.

workspaceDescription_dataSources :: Lens' WorkspaceDescription [DataSourceType]

Specifies the Amazon Web Services data sources that have been configured to have IAM roles and permissions created to allow Amazon Managed Grafana to read data from these sources.

workspaceDescription_endpoint :: Lens' WorkspaceDescription Text

The URL that users can use to access the Grafana console in the workspace.

workspaceDescription_grafanaVersion :: Lens' WorkspaceDescription Text

The version of Grafana supported in this workspace.

workspaceDescription_id :: Lens' WorkspaceDescription Text

The unique ID of this workspace.

workspaceDescription_modified :: Lens' WorkspaceDescription UTCTime

The most recent date that the workspace was modified.

workspaceDescription_status :: Lens' WorkspaceDescription WorkspaceStatus

The current status of the workspace.

WorkspaceSummary

data WorkspaceSummary

A structure that contains some information about one workspace in the account.

See: newWorkspaceSummary smart constructor.

Constructors

WorkspaceSummary'

Fields

name :: Maybe (Sensitive Text)

The name of the workspace.

notificationDestinations :: Maybe [NotificationDestinationType]

The Amazon Web Services notification channels that Amazon Managed Grafana can automatically create IAM roles and permissions for, which allows Amazon Managed Grafana to use these channels.

description :: Maybe (Sensitive Text)

The customer-entered description of the workspace.

authentication :: AuthenticationSummary

A structure containing information about the authentication methods used in the workspace.

created :: POSIX

The date that the workspace was created.

endpoint :: Text

The URL endpoint to use to access the Grafana console in the workspace.

grafanaVersion :: Text

The Grafana version that the workspace is running.

id :: Text

The unique ID of the workspace.

modified :: POSIX

The most recent date that the workspace was modified.

status :: WorkspaceStatus

The current status of the workspace.

newWorkspaceSummary

Arguments

:: AuthenticationSummary
$sel:authentication:WorkspaceSummary'

-> UTCTime
$sel:created:WorkspaceSummary'

-> Text
$sel:endpoint:WorkspaceSummary'

-> Text
$sel:grafanaVersion:WorkspaceSummary'

-> Text
$sel:id:WorkspaceSummary'

-> UTCTime
$sel:modified:WorkspaceSummary'

-> WorkspaceStatus
$sel:status:WorkspaceSummary'

-> WorkspaceSummary

Create a value of WorkspaceSummary with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:name:WorkspaceSummary', workspaceSummary_name - The name of the workspace.

$sel:notificationDestinations:WorkspaceSummary', workspaceSummary_notificationDestinations - The Amazon Web Services notification channels that Amazon Managed Grafana can automatically create IAM roles and permissions for, which allows Amazon Managed Grafana to use these channels.

$sel:description:WorkspaceSummary', workspaceSummary_description - The customer-entered description of the workspace.

$sel:authentication:WorkspaceSummary', workspaceSummary_authentication - A structure containing information about the authentication methods used in the workspace.

$sel:created:WorkspaceSummary', workspaceSummary_created - The date that the workspace was created.

$sel:endpoint:WorkspaceSummary', workspaceSummary_endpoint - The URL endpoint to use to access the Grafana console in the workspace.

$sel:grafanaVersion:WorkspaceSummary', workspaceSummary_grafanaVersion - The Grafana version that the workspace is running.

$sel:id:WorkspaceSummary', workspaceSummary_id - The unique ID of the workspace.

$sel:modified:WorkspaceSummary', workspaceSummary_modified - The most recent date that the workspace was modified.

$sel:status:WorkspaceSummary', workspaceSummary_status - The current status of the workspace.

workspaceSummary_name :: Lens' WorkspaceSummary (Maybe Text)

The name of the workspace.

workspaceSummary_notificationDestinations :: Lens' WorkspaceSummary (Maybe [NotificationDestinationType])

The Amazon Web Services notification channels that Amazon Managed Grafana can automatically create IAM roles and permissions for, which allows Amazon Managed Grafana to use these channels.

workspaceSummary_description :: Lens' WorkspaceSummary (Maybe Text)

The customer-entered description of the workspace.

workspaceSummary_authentication :: Lens' WorkspaceSummary AuthenticationSummary

A structure containing information about the authentication methods used in the workspace.

workspaceSummary_created :: Lens' WorkspaceSummary UTCTime

The date that the workspace was created.

workspaceSummary_endpoint :: Lens' WorkspaceSummary Text

The URL endpoint to use to access the Grafana console in the workspace.

workspaceSummary_grafanaVersion :: Lens' WorkspaceSummary Text

The Grafana version that the workspace is running.

workspaceSummary_id :: Lens' WorkspaceSummary Text

The unique ID of the workspace.

workspaceSummary_modified :: Lens' WorkspaceSummary UTCTime

The most recent date that the workspace was modified.

workspaceSummary_status :: Lens' WorkspaceSummary WorkspaceStatus

The current status of the workspace.