Amazonka.Grafana.Types.AuthenticationDescription

Description

Documentation

data AuthenticationDescription

A structure containing information about the user authentication methods used by the workspace.

See: newAuthenticationDescription smart constructor.

Constructors

AuthenticationDescription'

Fields

awsSso :: Maybe AwsSsoAuthentication

A structure containing information about how this workspace works with Amazon Web Services SSO.

saml :: Maybe SamlAuthentication

A structure containing information about how this workspace works with SAML, including what attributes within the assertion are to be mapped to user information in the workspace.

providers :: [AuthenticationProviderTypes]

Specifies whether this workspace uses Amazon Web Services SSO, SAML, or both methods to authenticate users to use the Grafana console in the Amazon Managed Grafana workspace.

newAuthenticationDescription :: AuthenticationDescription

Create a value of AuthenticationDescription with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:awsSso:AuthenticationDescription', authenticationDescription_awsSso - A structure containing information about how this workspace works with Amazon Web Services SSO.

$sel:saml:AuthenticationDescription', authenticationDescription_saml - A structure containing information about how this workspace works with SAML, including what attributes within the assertion are to be mapped to user information in the workspace.

$sel:providers:AuthenticationDescription', authenticationDescription_providers - Specifies whether this workspace uses Amazon Web Services SSO, SAML, or both methods to authenticate users to use the Grafana console in the Amazon Managed Grafana workspace.

authenticationDescription_awsSso :: Lens' AuthenticationDescription (Maybe AwsSsoAuthentication)

A structure containing information about how this workspace works with Amazon Web Services SSO.

authenticationDescription_saml :: Lens' AuthenticationDescription (Maybe SamlAuthentication)

A structure containing information about how this workspace works with SAML, including what attributes within the assertion are to be mapped to user information in the workspace.

authenticationDescription_providers :: Lens' AuthenticationDescription [AuthenticationProviderTypes]

Specifies whether this workspace uses Amazon Web Services SSO, SAML, or both methods to authenticate users to use the Grafana console in the Amazon Managed Grafana workspace.