Amazonka.Grafana.Types.UpdateError

Description

Documentation

data UpdateError

A structure containing information about one error encountered while performing an UpdatePermissions operation.

See: newUpdateError smart constructor.

Constructors

UpdateError'

Fields

causedBy :: UpdateInstruction

Specifies which permission update caused the error.

code :: Natural

The error code.

message :: Text

The message for this error.

newUpdateError

Arguments

:: UpdateInstruction
$sel:causedBy:UpdateError'

-> Natural
$sel:code:UpdateError'

-> Text
$sel:message:UpdateError'

-> UpdateError

Create a value of UpdateError with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:causedBy:UpdateError', updateError_causedBy - Specifies which permission update caused the error.

$sel:code:UpdateError', updateError_code - The error code.

$sel:message:UpdateError', updateError_message - The message for this error.

updateError_causedBy :: Lens' UpdateError UpdateInstruction

Specifies which permission update caused the error.

updateError_code :: Lens' UpdateError Natural

The error code.

updateError_message :: Lens' UpdateError Text

The message for this error.