Amazonka.Grafana.Types.WorkspaceSummary

Description

Documentation

data WorkspaceSummary

A structure that contains some information about one workspace in the account.

See: newWorkspaceSummary smart constructor.

Constructors

WorkspaceSummary'

Fields

name :: Maybe (Sensitive Text)

The name of the workspace.

notificationDestinations :: Maybe [NotificationDestinationType]

The Amazon Web Services notification channels that Amazon Managed Grafana can automatically create IAM roles and permissions for, which allows Amazon Managed Grafana to use these channels.

description :: Maybe (Sensitive Text)

The customer-entered description of the workspace.

authentication :: AuthenticationSummary

A structure containing information about the authentication methods used in the workspace.

created :: POSIX

The date that the workspace was created.

endpoint :: Text

The URL endpoint to use to access the Grafana console in the workspace.

grafanaVersion :: Text

The Grafana version that the workspace is running.

id :: Text

The unique ID of the workspace.

modified :: POSIX

The most recent date that the workspace was modified.

status :: WorkspaceStatus

The current status of the workspace.

newWorkspaceSummary

Arguments

:: AuthenticationSummary
$sel:authentication:WorkspaceSummary'

-> UTCTime
$sel:created:WorkspaceSummary'

-> Text
$sel:endpoint:WorkspaceSummary'

-> Text
$sel:grafanaVersion:WorkspaceSummary'

-> Text
$sel:id:WorkspaceSummary'

-> UTCTime
$sel:modified:WorkspaceSummary'

-> WorkspaceStatus
$sel:status:WorkspaceSummary'

-> WorkspaceSummary

Create a value of WorkspaceSummary with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:name:WorkspaceSummary', workspaceSummary_name - The name of the workspace.

$sel:notificationDestinations:WorkspaceSummary', workspaceSummary_notificationDestinations - The Amazon Web Services notification channels that Amazon Managed Grafana can automatically create IAM roles and permissions for, which allows Amazon Managed Grafana to use these channels.

$sel:description:WorkspaceSummary', workspaceSummary_description - The customer-entered description of the workspace.

$sel:authentication:WorkspaceSummary', workspaceSummary_authentication - A structure containing information about the authentication methods used in the workspace.

$sel:created:WorkspaceSummary', workspaceSummary_created - The date that the workspace was created.

$sel:endpoint:WorkspaceSummary', workspaceSummary_endpoint - The URL endpoint to use to access the Grafana console in the workspace.

$sel:grafanaVersion:WorkspaceSummary', workspaceSummary_grafanaVersion - The Grafana version that the workspace is running.

$sel:id:WorkspaceSummary', workspaceSummary_id - The unique ID of the workspace.

$sel:modified:WorkspaceSummary', workspaceSummary_modified - The most recent date that the workspace was modified.

$sel:status:WorkspaceSummary', workspaceSummary_status - The current status of the workspace.

workspaceSummary_name :: Lens' WorkspaceSummary (Maybe Text)

The name of the workspace.

workspaceSummary_notificationDestinations :: Lens' WorkspaceSummary (Maybe [NotificationDestinationType])

The Amazon Web Services notification channels that Amazon Managed Grafana can automatically create IAM roles and permissions for, which allows Amazon Managed Grafana to use these channels.

workspaceSummary_description :: Lens' WorkspaceSummary (Maybe Text)

The customer-entered description of the workspace.

workspaceSummary_authentication :: Lens' WorkspaceSummary AuthenticationSummary

A structure containing information about the authentication methods used in the workspace.

workspaceSummary_created :: Lens' WorkspaceSummary UTCTime

The date that the workspace was created.

workspaceSummary_endpoint :: Lens' WorkspaceSummary Text

The URL endpoint to use to access the Grafana console in the workspace.

workspaceSummary_grafanaVersion :: Lens' WorkspaceSummary Text

The Grafana version that the workspace is running.

workspaceSummary_id :: Lens' WorkspaceSummary Text

The unique ID of the workspace.

workspaceSummary_modified :: Lens' WorkspaceSummary UTCTime

The most recent date that the workspace was modified.

workspaceSummary_status :: Lens' WorkspaceSummary WorkspaceStatus

The current status of the workspace.