Amazonka.Grafana.UpdateWorkspace

Description

Modifies an existing Amazon Managed Grafana workspace. If you use this operation and omit any optional parameters, the existing values of those parameters are not changed.

To modify the user authentication methods that the workspace uses, such as SAML or Amazon Web Services SSO, use UpdateWorkspaceAuthentication.

To modify which users in the workspace have the Admin and Editor Grafana roles, use UpdatePermissions.

Creating a Request

data UpdateWorkspace

See: newUpdateWorkspace smart constructor.

Constructors

UpdateWorkspace'

Fields

workspaceRoleArn :: Maybe (Sensitive Text)

The workspace needs an IAM role that grants permissions to the Amazon Web Services resources that the workspace will view data from. If you already have a role that you want to use, specify it here. If you omit this field and you specify some Amazon Web Services resources in workspaceDataSources or workspaceNotificationDestinations, a new IAM role with the necessary permissions is automatically created.

workspaceDataSources :: Maybe [DataSourceType]

Specify the Amazon Web Services data sources that you want to be queried in this workspace. Specifying these data sources here enables Amazon Managed Grafana to create IAM roles and permissions that allow Amazon Managed Grafana to read data from these sources. You must still add them as data sources in the Grafana console in the workspace.

If you don't specify a data source here, you can still add it as a data source later in the workspace console. However, you will then have to manually configure permissions for it.

permissionType :: Maybe PermissionType

If you specify Service Managed, Amazon Managed Grafana automatically creates the IAM roles and provisions the permissions that the workspace needs to use Amazon Web Services data sources and notification channels.

If you specify CUSTOMER_MANAGED, you will manage those roles and permissions yourself. If you are creating this workspace in a member account of an organization and that account is not a delegated administrator account, and you want the workspace to access data sources in other Amazon Web Services accounts in the organization, you must choose CUSTOMER_MANAGED.

For more information, see Amazon Managed Grafana permissions and policies for Amazon Web Services data sources and notification channels

workspaceNotificationDestinations :: Maybe [NotificationDestinationType]

Specify the Amazon Web Services notification channels that you plan to use in this workspace. Specifying these data sources here enables Amazon Managed Grafana to create IAM roles and permissions that allow Amazon Managed Grafana to use these channels.

workspaceName :: Maybe (Sensitive Text)

A new name for the workspace to update.

accountAccessType :: Maybe AccountAccessType

Specifies whether the workspace can access Amazon Web Services resources in this Amazon Web Services account only, or whether it can also access Amazon Web Services resources in other accounts in the same organization. If you specify ORGANIZATION, you must specify which organizational units the workspace can access in the workspaceOrganizationalUnits parameter.

organizationRoleName :: Maybe (Sensitive Text)

The name of an IAM role that already exists to use to access resources through Organizations.

workspaceDescription :: Maybe (Sensitive Text)

A description for the workspace. This is used only to help you identify this workspace.

stackSetName :: Maybe Text

The name of the CloudFormation stack set to use to generate IAM roles to be used for this workspace.

workspaceOrganizationalUnits :: Maybe (Sensitive [Text])

Specifies the organizational units that this workspace is allowed to use data sources from, if this workspace is in an account that is part of an organization.

workspaceId :: Text

The ID of the workspace to update.

newUpdateWorkspace

Arguments

:: Text
$sel:workspaceId:UpdateWorkspace'

-> UpdateWorkspace

Create a value of UpdateWorkspace with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:workspaceRoleArn:UpdateWorkspace', updateWorkspace_workspaceRoleArn - The workspace needs an IAM role that grants permissions to the Amazon Web Services resources that the workspace will view data from. If you already have a role that you want to use, specify it here. If you omit this field and you specify some Amazon Web Services resources in workspaceDataSources or workspaceNotificationDestinations, a new IAM role with the necessary permissions is automatically created.

$sel:workspaceDataSources:UpdateWorkspace', updateWorkspace_workspaceDataSources - Specify the Amazon Web Services data sources that you want to be queried in this workspace. Specifying these data sources here enables Amazon Managed Grafana to create IAM roles and permissions that allow Amazon Managed Grafana to read data from these sources. You must still add them as data sources in the Grafana console in the workspace.

If you don't specify a data source here, you can still add it as a data source later in the workspace console. However, you will then have to manually configure permissions for it.

$sel:permissionType:UpdateWorkspace', updateWorkspace_permissionType - If you specify Service Managed, Amazon Managed Grafana automatically creates the IAM roles and provisions the permissions that the workspace needs to use Amazon Web Services data sources and notification channels.

If you specify CUSTOMER_MANAGED, you will manage those roles and permissions yourself. If you are creating this workspace in a member account of an organization and that account is not a delegated administrator account, and you want the workspace to access data sources in other Amazon Web Services accounts in the organization, you must choose CUSTOMER_MANAGED.

For more information, see Amazon Managed Grafana permissions and policies for Amazon Web Services data sources and notification channels

$sel:workspaceNotificationDestinations:UpdateWorkspace', updateWorkspace_workspaceNotificationDestinations - Specify the Amazon Web Services notification channels that you plan to use in this workspace. Specifying these data sources here enables Amazon Managed Grafana to create IAM roles and permissions that allow Amazon Managed Grafana to use these channels.

$sel:workspaceName:UpdateWorkspace', updateWorkspace_workspaceName - A new name for the workspace to update.

$sel:accountAccessType:UpdateWorkspace', updateWorkspace_accountAccessType - Specifies whether the workspace can access Amazon Web Services resources in this Amazon Web Services account only, or whether it can also access Amazon Web Services resources in other accounts in the same organization. If you specify ORGANIZATION, you must specify which organizational units the workspace can access in the workspaceOrganizationalUnits parameter.

$sel:organizationRoleName:UpdateWorkspace', updateWorkspace_organizationRoleName - The name of an IAM role that already exists to use to access resources through Organizations.

$sel:workspaceDescription:UpdateWorkspace', updateWorkspace_workspaceDescription - A description for the workspace. This is used only to help you identify this workspace.

$sel:stackSetName:UpdateWorkspace', updateWorkspace_stackSetName - The name of the CloudFormation stack set to use to generate IAM roles to be used for this workspace.

$sel:workspaceOrganizationalUnits:UpdateWorkspace', updateWorkspace_workspaceOrganizationalUnits - Specifies the organizational units that this workspace is allowed to use data sources from, if this workspace is in an account that is part of an organization.

$sel:workspaceId:UpdateWorkspace', updateWorkspace_workspaceId - The ID of the workspace to update.

Request Lenses

updateWorkspace_workspaceRoleArn :: Lens' UpdateWorkspace (Maybe Text)

The workspace needs an IAM role that grants permissions to the Amazon Web Services resources that the workspace will view data from. If you already have a role that you want to use, specify it here. If you omit this field and you specify some Amazon Web Services resources in workspaceDataSources or workspaceNotificationDestinations, a new IAM role with the necessary permissions is automatically created.

updateWorkspace_workspaceDataSources :: Lens' UpdateWorkspace (Maybe [DataSourceType])

Specify the Amazon Web Services data sources that you want to be queried in this workspace. Specifying these data sources here enables Amazon Managed Grafana to create IAM roles and permissions that allow Amazon Managed Grafana to read data from these sources. You must still add them as data sources in the Grafana console in the workspace.

If you don't specify a data source here, you can still add it as a data source later in the workspace console. However, you will then have to manually configure permissions for it.

updateWorkspace_permissionType :: Lens' UpdateWorkspace (Maybe PermissionType)

If you specify Service Managed, Amazon Managed Grafana automatically creates the IAM roles and provisions the permissions that the workspace needs to use Amazon Web Services data sources and notification channels.

If you specify CUSTOMER_MANAGED, you will manage those roles and permissions yourself. If you are creating this workspace in a member account of an organization and that account is not a delegated administrator account, and you want the workspace to access data sources in other Amazon Web Services accounts in the organization, you must choose CUSTOMER_MANAGED.

For more information, see Amazon Managed Grafana permissions and policies for Amazon Web Services data sources and notification channels

updateWorkspace_workspaceNotificationDestinations :: Lens' UpdateWorkspace (Maybe [NotificationDestinationType])

Specify the Amazon Web Services notification channels that you plan to use in this workspace. Specifying these data sources here enables Amazon Managed Grafana to create IAM roles and permissions that allow Amazon Managed Grafana to use these channels.

updateWorkspace_workspaceName :: Lens' UpdateWorkspace (Maybe Text)

A new name for the workspace to update.

updateWorkspace_accountAccessType :: Lens' UpdateWorkspace (Maybe AccountAccessType)

Specifies whether the workspace can access Amazon Web Services resources in this Amazon Web Services account only, or whether it can also access Amazon Web Services resources in other accounts in the same organization. If you specify ORGANIZATION, you must specify which organizational units the workspace can access in the workspaceOrganizationalUnits parameter.

updateWorkspace_organizationRoleName :: Lens' UpdateWorkspace (Maybe Text)

The name of an IAM role that already exists to use to access resources through Organizations.

updateWorkspace_workspaceDescription :: Lens' UpdateWorkspace (Maybe Text)

A description for the workspace. This is used only to help you identify this workspace.

updateWorkspace_stackSetName :: Lens' UpdateWorkspace (Maybe Text)

The name of the CloudFormation stack set to use to generate IAM roles to be used for this workspace.

updateWorkspace_workspaceOrganizationalUnits :: Lens' UpdateWorkspace (Maybe [Text])

Specifies the organizational units that this workspace is allowed to use data sources from, if this workspace is in an account that is part of an organization.

updateWorkspace_workspaceId :: Lens' UpdateWorkspace Text

The ID of the workspace to update.

Destructuring the Response

data UpdateWorkspaceResponse

See: newUpdateWorkspaceResponse smart constructor.

Constructors

UpdateWorkspaceResponse'

Fields

httpStatus :: Int

The response's http status code.

workspace :: WorkspaceDescription

A structure containing data about the workspace that was created.

newUpdateWorkspaceResponse

Arguments

:: Int
$sel:httpStatus:UpdateWorkspaceResponse'

-> WorkspaceDescription
$sel:workspace:UpdateWorkspaceResponse'

-> UpdateWorkspaceResponse

Create a value of UpdateWorkspaceResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:httpStatus:UpdateWorkspaceResponse', updateWorkspaceResponse_httpStatus - The response's http status code.

$sel:workspace:UpdateWorkspaceResponse', updateWorkspaceResponse_workspace - A structure containing data about the workspace that was created.

Response Lenses

updateWorkspaceResponse_httpStatus :: Lens' UpdateWorkspaceResponse Int

The response's http status code.

updateWorkspaceResponse_workspace :: Lens' UpdateWorkspaceResponse WorkspaceDescription

A structure containing data about the workspace that was created.