Amazonka.Grafana.UpdateWorkspaceAuthentication

Description

Use this operation to define the identity provider (IdP) that this workspace authenticates users from, using SAML. You can also map SAML assertion attributes to workspace user information and define which groups in the assertion attribute are to have the Admin and Editor roles in the workspace.

Creating a Request

data UpdateWorkspaceAuthentication

See: newUpdateWorkspaceAuthentication smart constructor.

Constructors

UpdateWorkspaceAuthentication'

Fields

samlConfiguration :: Maybe SamlConfiguration

If the workspace uses SAML, use this structure to map SAML assertion attributes to workspace user information and define which groups in the assertion attribute are to have the Admin and Editor roles in the workspace.

authenticationProviders :: [AuthenticationProviderTypes]

Specifies whether this workspace uses SAML 2.0, Amazon Web Services Single Sign On, or both to authenticate users for using the Grafana console within a workspace. For more information, see User authentication in Amazon Managed Grafana.

workspaceId :: Text

The ID of the workspace to update the authentication for.

newUpdateWorkspaceAuthentication

Arguments

:: Text
$sel:workspaceId:UpdateWorkspaceAuthentication'

-> UpdateWorkspaceAuthentication

Create a value of UpdateWorkspaceAuthentication with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:samlConfiguration:UpdateWorkspaceAuthentication', updateWorkspaceAuthentication_samlConfiguration - If the workspace uses SAML, use this structure to map SAML assertion attributes to workspace user information and define which groups in the assertion attribute are to have the Admin and Editor roles in the workspace.

$sel:authenticationProviders:UpdateWorkspaceAuthentication', updateWorkspaceAuthentication_authenticationProviders - Specifies whether this workspace uses SAML 2.0, Amazon Web Services Single Sign On, or both to authenticate users for using the Grafana console within a workspace. For more information, see User authentication in Amazon Managed Grafana.

$sel:workspaceId:UpdateWorkspaceAuthentication', updateWorkspaceAuthentication_workspaceId - The ID of the workspace to update the authentication for.

Request Lenses

updateWorkspaceAuthentication_samlConfiguration :: Lens' UpdateWorkspaceAuthentication (Maybe SamlConfiguration)

If the workspace uses SAML, use this structure to map SAML assertion attributes to workspace user information and define which groups in the assertion attribute are to have the Admin and Editor roles in the workspace.

updateWorkspaceAuthentication_authenticationProviders :: Lens' UpdateWorkspaceAuthentication [AuthenticationProviderTypes]

Specifies whether this workspace uses SAML 2.0, Amazon Web Services Single Sign On, or both to authenticate users for using the Grafana console within a workspace. For more information, see User authentication in Amazon Managed Grafana.

updateWorkspaceAuthentication_workspaceId :: Lens' UpdateWorkspaceAuthentication Text

The ID of the workspace to update the authentication for.

Destructuring the Response

data UpdateWorkspaceAuthenticationResponse

See: newUpdateWorkspaceAuthenticationResponse smart constructor.

Constructors

UpdateWorkspaceAuthenticationResponse'

Fields

httpStatus :: Int

The response's http status code.

authentication :: AuthenticationDescription

A structure that describes the user authentication for this workspace after the update is made.

newUpdateWorkspaceAuthenticationResponse

Arguments

:: Int
$sel:httpStatus:UpdateWorkspaceAuthenticationResponse'

-> AuthenticationDescription
$sel:authentication:UpdateWorkspaceAuthenticationResponse'

-> UpdateWorkspaceAuthenticationResponse

Create a value of UpdateWorkspaceAuthenticationResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:httpStatus:UpdateWorkspaceAuthenticationResponse', updateWorkspaceAuthenticationResponse_httpStatus - The response's http status code.

$sel:authentication:UpdateWorkspaceAuthenticationResponse', updateWorkspaceAuthenticationResponse_authentication - A structure that describes the user authentication for this workspace after the update is made.

Response Lenses

updateWorkspaceAuthenticationResponse_httpStatus :: Lens' UpdateWorkspaceAuthenticationResponse Int

The response's http status code.

updateWorkspaceAuthenticationResponse_authentication :: Lens' UpdateWorkspaceAuthenticationResponse AuthenticationDescription

A structure that describes the user authentication for this workspace after the update is made.