Amazonka.Greengrass.CreateConnectorDefinition

Description

Creates a connector definition. You may provide the initial version of the connector definition now or use ''CreateConnectorDefinitionVersion'' at a later time.

Creating a Request

data CreateConnectorDefinition

See: newCreateConnectorDefinition smart constructor.

Constructors

CreateConnectorDefinition'

Fields

amznClientToken :: Maybe Text

A client token used to correlate requests and responses.

initialVersion :: Maybe ConnectorDefinitionVersion

Information about the initial version of the connector definition.

name :: Maybe Text

The name of the connector definition.

tags :: Maybe (HashMap Text Text)

Tag(s) to add to the new resource.

newCreateConnectorDefinition :: CreateConnectorDefinition

Create a value of CreateConnectorDefinition with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:amznClientToken:CreateConnectorDefinition', createConnectorDefinition_amznClientToken - A client token used to correlate requests and responses.

$sel:initialVersion:CreateConnectorDefinition', createConnectorDefinition_initialVersion - Information about the initial version of the connector definition.

$sel:name:CreateConnectorDefinition', createConnectorDefinition_name - The name of the connector definition.

$sel:tags:CreateConnectorDefinition', createConnectorDefinition_tags - Tag(s) to add to the new resource.

Request Lenses

createConnectorDefinition_amznClientToken :: Lens' CreateConnectorDefinition (Maybe Text)

A client token used to correlate requests and responses.

createConnectorDefinition_initialVersion :: Lens' CreateConnectorDefinition (Maybe ConnectorDefinitionVersion)

Information about the initial version of the connector definition.

createConnectorDefinition_name :: Lens' CreateConnectorDefinition (Maybe Text)

The name of the connector definition.

createConnectorDefinition_tags :: Lens' CreateConnectorDefinition (Maybe (HashMap Text Text))

Tag(s) to add to the new resource.

Destructuring the Response

data CreateConnectorDefinitionResponse

See: newCreateConnectorDefinitionResponse smart constructor.

Constructors

CreateConnectorDefinitionResponse'

Fields

latestVersionArn :: Maybe Text

The ARN of the latest version associated with the definition.

arn :: Maybe Text

The ARN of the definition.

name :: Maybe Text

The name of the definition.

creationTimestamp :: Maybe Text

The time, in milliseconds since the epoch, when the definition was created.

id :: Maybe Text

The ID of the definition.

latestVersion :: Maybe Text

The ID of the latest version associated with the definition.

lastUpdatedTimestamp :: Maybe Text

The time, in milliseconds since the epoch, when the definition was last updated.

httpStatus :: Int

The response's http status code.

newCreateConnectorDefinitionResponse

Arguments

:: Int
$sel:httpStatus:CreateConnectorDefinitionResponse'

-> CreateConnectorDefinitionResponse

Create a value of CreateConnectorDefinitionResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:latestVersionArn:CreateConnectorDefinitionResponse', createConnectorDefinitionResponse_latestVersionArn - The ARN of the latest version associated with the definition.

$sel:arn:CreateConnectorDefinitionResponse', createConnectorDefinitionResponse_arn - The ARN of the definition.

$sel:name:CreateConnectorDefinition', createConnectorDefinitionResponse_name - The name of the definition.

$sel:creationTimestamp:CreateConnectorDefinitionResponse', createConnectorDefinitionResponse_creationTimestamp - The time, in milliseconds since the epoch, when the definition was created.

$sel:id:CreateConnectorDefinitionResponse', createConnectorDefinitionResponse_id - The ID of the definition.

$sel:latestVersion:CreateConnectorDefinitionResponse', createConnectorDefinitionResponse_latestVersion - The ID of the latest version associated with the definition.

$sel:lastUpdatedTimestamp:CreateConnectorDefinitionResponse', createConnectorDefinitionResponse_lastUpdatedTimestamp - The time, in milliseconds since the epoch, when the definition was last updated.

$sel:httpStatus:CreateConnectorDefinitionResponse', createConnectorDefinitionResponse_httpStatus - The response's http status code.

Response Lenses

createConnectorDefinitionResponse_latestVersionArn :: Lens' CreateConnectorDefinitionResponse (Maybe Text)

The ARN of the latest version associated with the definition.

createConnectorDefinitionResponse_arn :: Lens' CreateConnectorDefinitionResponse (Maybe Text)

The ARN of the definition.

createConnectorDefinitionResponse_name :: Lens' CreateConnectorDefinitionResponse (Maybe Text)

The name of the definition.

createConnectorDefinitionResponse_creationTimestamp :: Lens' CreateConnectorDefinitionResponse (Maybe Text)

The time, in milliseconds since the epoch, when the definition was created.

createConnectorDefinitionResponse_id :: Lens' CreateConnectorDefinitionResponse (Maybe Text)

The ID of the definition.

createConnectorDefinitionResponse_latestVersion :: Lens' CreateConnectorDefinitionResponse (Maybe Text)

The ID of the latest version associated with the definition.

createConnectorDefinitionResponse_lastUpdatedTimestamp :: Lens' CreateConnectorDefinitionResponse (Maybe Text)

The time, in milliseconds since the epoch, when the definition was last updated.

createConnectorDefinitionResponse_httpStatus :: Lens' CreateConnectorDefinitionResponse Int

The response's http status code.