Amazonka.Greengrass.CreateDeviceDefinitionVersion

Description

Creates a version of a device definition that has already been defined.

Creating a Request

data CreateDeviceDefinitionVersion

See: newCreateDeviceDefinitionVersion smart constructor.

Constructors

CreateDeviceDefinitionVersion'

Fields

amznClientToken :: Maybe Text

A client token used to correlate requests and responses.

devices :: Maybe [Device]

A list of devices in the definition version.

deviceDefinitionId :: Text

The ID of the device definition.

newCreateDeviceDefinitionVersion

Arguments

:: Text
$sel:deviceDefinitionId:CreateDeviceDefinitionVersion'

-> CreateDeviceDefinitionVersion

Create a value of CreateDeviceDefinitionVersion with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:amznClientToken:CreateDeviceDefinitionVersion', createDeviceDefinitionVersion_amznClientToken - A client token used to correlate requests and responses.

$sel:devices:CreateDeviceDefinitionVersion', createDeviceDefinitionVersion_devices - A list of devices in the definition version.

$sel:deviceDefinitionId:CreateDeviceDefinitionVersion', createDeviceDefinitionVersion_deviceDefinitionId - The ID of the device definition.

Request Lenses

createDeviceDefinitionVersion_amznClientToken :: Lens' CreateDeviceDefinitionVersion (Maybe Text)

A client token used to correlate requests and responses.

createDeviceDefinitionVersion_devices :: Lens' CreateDeviceDefinitionVersion (Maybe [Device])

A list of devices in the definition version.

createDeviceDefinitionVersion_deviceDefinitionId :: Lens' CreateDeviceDefinitionVersion Text

The ID of the device definition.

Destructuring the Response

data CreateDeviceDefinitionVersionResponse

See: newCreateDeviceDefinitionVersionResponse smart constructor.

Constructors

CreateDeviceDefinitionVersionResponse'

Fields

arn :: Maybe Text

The ARN of the version.

creationTimestamp :: Maybe Text

The time, in milliseconds since the epoch, when the version was created.

version :: Maybe Text

The ID of the version.

id :: Maybe Text

The ID of the parent definition that the version is associated with.

httpStatus :: Int

The response's http status code.

newCreateDeviceDefinitionVersionResponse

Arguments

:: Int
$sel:httpStatus:CreateDeviceDefinitionVersionResponse'

-> CreateDeviceDefinitionVersionResponse

Create a value of CreateDeviceDefinitionVersionResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:arn:CreateDeviceDefinitionVersionResponse', createDeviceDefinitionVersionResponse_arn - The ARN of the version.

$sel:creationTimestamp:CreateDeviceDefinitionVersionResponse', createDeviceDefinitionVersionResponse_creationTimestamp - The time, in milliseconds since the epoch, when the version was created.

$sel:version:CreateDeviceDefinitionVersionResponse', createDeviceDefinitionVersionResponse_version - The ID of the version.

$sel:id:CreateDeviceDefinitionVersionResponse', createDeviceDefinitionVersionResponse_id - The ID of the parent definition that the version is associated with.

$sel:httpStatus:CreateDeviceDefinitionVersionResponse', createDeviceDefinitionVersionResponse_httpStatus - The response's http status code.

Response Lenses

createDeviceDefinitionVersionResponse_arn :: Lens' CreateDeviceDefinitionVersionResponse (Maybe Text)

The ARN of the version.

createDeviceDefinitionVersionResponse_creationTimestamp :: Lens' CreateDeviceDefinitionVersionResponse (Maybe Text)

The time, in milliseconds since the epoch, when the version was created.

createDeviceDefinitionVersionResponse_version :: Lens' CreateDeviceDefinitionVersionResponse (Maybe Text)

The ID of the version.

createDeviceDefinitionVersionResponse_id :: Lens' CreateDeviceDefinitionVersionResponse (Maybe Text)

The ID of the parent definition that the version is associated with.

createDeviceDefinitionVersionResponse_httpStatus :: Lens' CreateDeviceDefinitionVersionResponse Int

The response's http status code.