Amazonka.Greengrass.CreateLoggerDefinition

Description

Creates a logger definition. You may provide the initial version of the logger definition now or use ''CreateLoggerDefinitionVersion'' at a later time.

Creating a Request

data CreateLoggerDefinition

See: newCreateLoggerDefinition smart constructor.

Constructors

CreateLoggerDefinition'

Fields

amznClientToken :: Maybe Text

A client token used to correlate requests and responses.

initialVersion :: Maybe LoggerDefinitionVersion

Information about the initial version of the logger definition.

name :: Maybe Text

The name of the logger definition.

tags :: Maybe (HashMap Text Text)

Tag(s) to add to the new resource.

newCreateLoggerDefinition :: CreateLoggerDefinition

Create a value of CreateLoggerDefinition with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:amznClientToken:CreateLoggerDefinition', createLoggerDefinition_amznClientToken - A client token used to correlate requests and responses.

$sel:initialVersion:CreateLoggerDefinition', createLoggerDefinition_initialVersion - Information about the initial version of the logger definition.

$sel:name:CreateLoggerDefinition', createLoggerDefinition_name - The name of the logger definition.

$sel:tags:CreateLoggerDefinition', createLoggerDefinition_tags - Tag(s) to add to the new resource.

Request Lenses

createLoggerDefinition_amznClientToken :: Lens' CreateLoggerDefinition (Maybe Text)

A client token used to correlate requests and responses.

createLoggerDefinition_initialVersion :: Lens' CreateLoggerDefinition (Maybe LoggerDefinitionVersion)

Information about the initial version of the logger definition.

createLoggerDefinition_name :: Lens' CreateLoggerDefinition (Maybe Text)

The name of the logger definition.

createLoggerDefinition_tags :: Lens' CreateLoggerDefinition (Maybe (HashMap Text Text))

Tag(s) to add to the new resource.

Destructuring the Response

data CreateLoggerDefinitionResponse

See: newCreateLoggerDefinitionResponse smart constructor.

Constructors

CreateLoggerDefinitionResponse'

Fields

latestVersionArn :: Maybe Text

The ARN of the latest version associated with the definition.

arn :: Maybe Text

The ARN of the definition.

name :: Maybe Text

The name of the definition.

creationTimestamp :: Maybe Text

The time, in milliseconds since the epoch, when the definition was created.

id :: Maybe Text

The ID of the definition.

latestVersion :: Maybe Text

The ID of the latest version associated with the definition.

lastUpdatedTimestamp :: Maybe Text

The time, in milliseconds since the epoch, when the definition was last updated.

httpStatus :: Int

The response's http status code.

newCreateLoggerDefinitionResponse

Arguments

:: Int
$sel:httpStatus:CreateLoggerDefinitionResponse'

-> CreateLoggerDefinitionResponse

Create a value of CreateLoggerDefinitionResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:latestVersionArn:CreateLoggerDefinitionResponse', createLoggerDefinitionResponse_latestVersionArn - The ARN of the latest version associated with the definition.

$sel:arn:CreateLoggerDefinitionResponse', createLoggerDefinitionResponse_arn - The ARN of the definition.

$sel:name:CreateLoggerDefinition', createLoggerDefinitionResponse_name - The name of the definition.

$sel:creationTimestamp:CreateLoggerDefinitionResponse', createLoggerDefinitionResponse_creationTimestamp - The time, in milliseconds since the epoch, when the definition was created.

$sel:id:CreateLoggerDefinitionResponse', createLoggerDefinitionResponse_id - The ID of the definition.

$sel:latestVersion:CreateLoggerDefinitionResponse', createLoggerDefinitionResponse_latestVersion - The ID of the latest version associated with the definition.

$sel:lastUpdatedTimestamp:CreateLoggerDefinitionResponse', createLoggerDefinitionResponse_lastUpdatedTimestamp - The time, in milliseconds since the epoch, when the definition was last updated.

$sel:httpStatus:CreateLoggerDefinitionResponse', createLoggerDefinitionResponse_httpStatus - The response's http status code.

Response Lenses

createLoggerDefinitionResponse_latestVersionArn :: Lens' CreateLoggerDefinitionResponse (Maybe Text)

The ARN of the latest version associated with the definition.

createLoggerDefinitionResponse_arn :: Lens' CreateLoggerDefinitionResponse (Maybe Text)

The ARN of the definition.

createLoggerDefinitionResponse_name :: Lens' CreateLoggerDefinitionResponse (Maybe Text)

The name of the definition.

createLoggerDefinitionResponse_creationTimestamp :: Lens' CreateLoggerDefinitionResponse (Maybe Text)

The time, in milliseconds since the epoch, when the definition was created.

createLoggerDefinitionResponse_id :: Lens' CreateLoggerDefinitionResponse (Maybe Text)

The ID of the definition.

createLoggerDefinitionResponse_latestVersion :: Lens' CreateLoggerDefinitionResponse (Maybe Text)

The ID of the latest version associated with the definition.

createLoggerDefinitionResponse_lastUpdatedTimestamp :: Lens' CreateLoggerDefinitionResponse (Maybe Text)

The time, in milliseconds since the epoch, when the definition was last updated.

createLoggerDefinitionResponse_httpStatus :: Lens' CreateLoggerDefinitionResponse Int

The response's http status code.