Amazonka.Greengrass.CreateSoftwareUpdateJob

Description

Creates a software update for a core or group of cores (specified as an IoT thing group.) Use this to update the OTA Agent as well as the Greengrass core software. It makes use of the IoT Jobs feature which provides additional commands to manage a Greengrass core software update job.

Creating a Request

data CreateSoftwareUpdateJob

See: newCreateSoftwareUpdateJob smart constructor.

Constructors

CreateSoftwareUpdateJob'

Fields

updateAgentLogLevel :: Maybe UpdateAgentLogLevel

amznClientToken :: Maybe Text

A client token used to correlate requests and responses.

s3UrlSignerRole :: Text

updateTargetsArchitecture :: UpdateTargetsArchitecture

softwareToUpdate :: SoftwareToUpdate

updateTargets :: [Text]

updateTargetsOperatingSystem :: UpdateTargetsOperatingSystem

newCreateSoftwareUpdateJob

Arguments

:: Text
$sel:s3UrlSignerRole:CreateSoftwareUpdateJob'

-> UpdateTargetsArchitecture
$sel:updateTargetsArchitecture:CreateSoftwareUpdateJob'

-> SoftwareToUpdate
$sel:softwareToUpdate:CreateSoftwareUpdateJob'

-> UpdateTargetsOperatingSystem
$sel:updateTargetsOperatingSystem:CreateSoftwareUpdateJob'

-> CreateSoftwareUpdateJob

Create a value of CreateSoftwareUpdateJob with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:updateAgentLogLevel:CreateSoftwareUpdateJob', createSoftwareUpdateJob_updateAgentLogLevel - Undocumented member.

$sel:amznClientToken:CreateSoftwareUpdateJob', createSoftwareUpdateJob_amznClientToken - A client token used to correlate requests and responses.

$sel:s3UrlSignerRole:CreateSoftwareUpdateJob', createSoftwareUpdateJob_s3UrlSignerRole - Undocumented member.

$sel:updateTargetsArchitecture:CreateSoftwareUpdateJob', createSoftwareUpdateJob_updateTargetsArchitecture - Undocumented member.

$sel:softwareToUpdate:CreateSoftwareUpdateJob', createSoftwareUpdateJob_softwareToUpdate - Undocumented member.

$sel:updateTargets:CreateSoftwareUpdateJob', createSoftwareUpdateJob_updateTargets - Undocumented member.

$sel:updateTargetsOperatingSystem:CreateSoftwareUpdateJob', createSoftwareUpdateJob_updateTargetsOperatingSystem - Undocumented member.

Request Lenses

createSoftwareUpdateJob_updateAgentLogLevel :: Lens' CreateSoftwareUpdateJob (Maybe UpdateAgentLogLevel)

Undocumented member.

createSoftwareUpdateJob_amznClientToken :: Lens' CreateSoftwareUpdateJob (Maybe Text)

A client token used to correlate requests and responses.

createSoftwareUpdateJob_s3UrlSignerRole :: Lens' CreateSoftwareUpdateJob Text

Undocumented member.

createSoftwareUpdateJob_updateTargetsArchitecture :: Lens' CreateSoftwareUpdateJob UpdateTargetsArchitecture

Undocumented member.

createSoftwareUpdateJob_softwareToUpdate :: Lens' CreateSoftwareUpdateJob SoftwareToUpdate

Undocumented member.

createSoftwareUpdateJob_updateTargets :: Lens' CreateSoftwareUpdateJob [Text]

Undocumented member.

createSoftwareUpdateJob_updateTargetsOperatingSystem :: Lens' CreateSoftwareUpdateJob UpdateTargetsOperatingSystem

Undocumented member.

Destructuring the Response

data CreateSoftwareUpdateJobResponse

See: newCreateSoftwareUpdateJobResponse smart constructor.

Constructors

CreateSoftwareUpdateJobResponse'

Fields

platformSoftwareVersion :: Maybe Text

The software version installed on the device or devices after the update.

iotJobArn :: Maybe Text

The IoT Job ARN corresponding to this update.

iotJobId :: Maybe Text

The IoT Job Id corresponding to this update.

httpStatus :: Int

The response's http status code.

newCreateSoftwareUpdateJobResponse

Arguments

:: Int
$sel:httpStatus:CreateSoftwareUpdateJobResponse'

-> CreateSoftwareUpdateJobResponse

Create a value of CreateSoftwareUpdateJobResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:platformSoftwareVersion:CreateSoftwareUpdateJobResponse', createSoftwareUpdateJobResponse_platformSoftwareVersion - The software version installed on the device or devices after the update.

$sel:iotJobArn:CreateSoftwareUpdateJobResponse', createSoftwareUpdateJobResponse_iotJobArn - The IoT Job ARN corresponding to this update.

$sel:iotJobId:CreateSoftwareUpdateJobResponse', createSoftwareUpdateJobResponse_iotJobId - The IoT Job Id corresponding to this update.

$sel:httpStatus:CreateSoftwareUpdateJobResponse', createSoftwareUpdateJobResponse_httpStatus - The response's http status code.

Response Lenses

createSoftwareUpdateJobResponse_platformSoftwareVersion :: Lens' CreateSoftwareUpdateJobResponse (Maybe Text)

The software version installed on the device or devices after the update.

createSoftwareUpdateJobResponse_iotJobArn :: Lens' CreateSoftwareUpdateJobResponse (Maybe Text)

The IoT Job ARN corresponding to this update.

createSoftwareUpdateJobResponse_iotJobId :: Lens' CreateSoftwareUpdateJobResponse (Maybe Text)

The IoT Job Id corresponding to this update.

createSoftwareUpdateJobResponse_httpStatus :: Lens' CreateSoftwareUpdateJobResponse Int

The response's http status code.