Amazonka.Greengrass.GetCoreDefinition

Description

Retrieves information about a core definition version.

Creating a Request

data GetCoreDefinition

See: newGetCoreDefinition smart constructor.

Constructors

GetCoreDefinition'

Fields

coreDefinitionId :: Text

The ID of the core definition.

newGetCoreDefinition

Arguments

:: Text
$sel:coreDefinitionId:GetCoreDefinition'

-> GetCoreDefinition

Create a value of GetCoreDefinition with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:coreDefinitionId:GetCoreDefinition', getCoreDefinition_coreDefinitionId - The ID of the core definition.

Request Lenses

getCoreDefinition_coreDefinitionId :: Lens' GetCoreDefinition Text

The ID of the core definition.

Destructuring the Response

data GetCoreDefinitionResponse

See: newGetCoreDefinitionResponse smart constructor.

Constructors

GetCoreDefinitionResponse'

Fields

latestVersionArn :: Maybe Text

The ARN of the latest version associated with the definition.

arn :: Maybe Text

The ARN of the definition.

name :: Maybe Text

The name of the definition.

creationTimestamp :: Maybe Text

The time, in milliseconds since the epoch, when the definition was created.

id :: Maybe Text

The ID of the definition.

latestVersion :: Maybe Text

The ID of the latest version associated with the definition.

lastUpdatedTimestamp :: Maybe Text

The time, in milliseconds since the epoch, when the definition was last updated.

tags :: Maybe (HashMap Text Text)

Tag(s) attached to the resource arn.

httpStatus :: Int

The response's http status code.

newGetCoreDefinitionResponse

Arguments

:: Int
$sel:httpStatus:GetCoreDefinitionResponse'

-> GetCoreDefinitionResponse

Create a value of GetCoreDefinitionResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:latestVersionArn:GetCoreDefinitionResponse', getCoreDefinitionResponse_latestVersionArn - The ARN of the latest version associated with the definition.

$sel:arn:GetCoreDefinitionResponse', getCoreDefinitionResponse_arn - The ARN of the definition.

$sel:name:GetCoreDefinitionResponse', getCoreDefinitionResponse_name - The name of the definition.

$sel:creationTimestamp:GetCoreDefinitionResponse', getCoreDefinitionResponse_creationTimestamp - The time, in milliseconds since the epoch, when the definition was created.

$sel:id:GetCoreDefinitionResponse', getCoreDefinitionResponse_id - The ID of the definition.

$sel:latestVersion:GetCoreDefinitionResponse', getCoreDefinitionResponse_latestVersion - The ID of the latest version associated with the definition.

$sel:lastUpdatedTimestamp:GetCoreDefinitionResponse', getCoreDefinitionResponse_lastUpdatedTimestamp - The time, in milliseconds since the epoch, when the definition was last updated.

$sel:tags:GetCoreDefinitionResponse', getCoreDefinitionResponse_tags - Tag(s) attached to the resource arn.

$sel:httpStatus:GetCoreDefinitionResponse', getCoreDefinitionResponse_httpStatus - The response's http status code.

Response Lenses

getCoreDefinitionResponse_latestVersionArn :: Lens' GetCoreDefinitionResponse (Maybe Text)

The ARN of the latest version associated with the definition.

getCoreDefinitionResponse_arn :: Lens' GetCoreDefinitionResponse (Maybe Text)

The ARN of the definition.

getCoreDefinitionResponse_name :: Lens' GetCoreDefinitionResponse (Maybe Text)

The name of the definition.

getCoreDefinitionResponse_creationTimestamp :: Lens' GetCoreDefinitionResponse (Maybe Text)

The time, in milliseconds since the epoch, when the definition was created.

getCoreDefinitionResponse_id :: Lens' GetCoreDefinitionResponse (Maybe Text)

The ID of the definition.

getCoreDefinitionResponse_latestVersion :: Lens' GetCoreDefinitionResponse (Maybe Text)

The ID of the latest version associated with the definition.

getCoreDefinitionResponse_lastUpdatedTimestamp :: Lens' GetCoreDefinitionResponse (Maybe Text)

The time, in milliseconds since the epoch, when the definition was last updated.

getCoreDefinitionResponse_tags :: Lens' GetCoreDefinitionResponse (Maybe (HashMap Text Text))

Tag(s) attached to the resource arn.

getCoreDefinitionResponse_httpStatus :: Lens' GetCoreDefinitionResponse Int

The response's http status code.