Amazonka.Greengrass.Types.BulkDeploymentResult

Description

Documentation

data BulkDeploymentResult

Information about an individual group deployment in a bulk deployment operation.

See: newBulkDeploymentResult smart constructor.

Constructors

BulkDeploymentResult'

Fields

deploymentId :: Maybe Text

The ID of the group deployment.

deploymentArn :: Maybe Text

The ARN of the group deployment.

createdAt :: Maybe Text

The time, in ISO format, when the deployment was created.

deploymentType :: Maybe DeploymentType

The type of the deployment.

errorDetails :: Maybe [ErrorDetail]

Details about the error.

groupArn :: Maybe Text

The ARN of the Greengrass group.

deploymentStatus :: Maybe Text

The current status of the group deployment: ''InProgress'', ''Building'', ''Success'', or ''Failure''.

errorMessage :: Maybe Text

The error message for a failed deployment

newBulkDeploymentResult :: BulkDeploymentResult

Create a value of BulkDeploymentResult with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:deploymentId:BulkDeploymentResult', bulkDeploymentResult_deploymentId - The ID of the group deployment.

$sel:deploymentArn:BulkDeploymentResult', bulkDeploymentResult_deploymentArn - The ARN of the group deployment.

$sel:createdAt:BulkDeploymentResult', bulkDeploymentResult_createdAt - The time, in ISO format, when the deployment was created.

$sel:deploymentType:BulkDeploymentResult', bulkDeploymentResult_deploymentType - The type of the deployment.

$sel:errorDetails:BulkDeploymentResult', bulkDeploymentResult_errorDetails - Details about the error.

$sel:groupArn:BulkDeploymentResult', bulkDeploymentResult_groupArn - The ARN of the Greengrass group.

$sel:deploymentStatus:BulkDeploymentResult', bulkDeploymentResult_deploymentStatus - The current status of the group deployment: ''InProgress'', ''Building'', ''Success'', or ''Failure''.

$sel:errorMessage:BulkDeploymentResult', bulkDeploymentResult_errorMessage - The error message for a failed deployment

bulkDeploymentResult_deploymentId :: Lens' BulkDeploymentResult (Maybe Text)

The ID of the group deployment.

bulkDeploymentResult_deploymentArn :: Lens' BulkDeploymentResult (Maybe Text)

The ARN of the group deployment.

bulkDeploymentResult_createdAt :: Lens' BulkDeploymentResult (Maybe Text)

The time, in ISO format, when the deployment was created.

bulkDeploymentResult_deploymentType :: Lens' BulkDeploymentResult (Maybe DeploymentType)

The type of the deployment.

bulkDeploymentResult_errorDetails :: Lens' BulkDeploymentResult (Maybe [ErrorDetail])

Details about the error.

bulkDeploymentResult_groupArn :: Lens' BulkDeploymentResult (Maybe Text)

The ARN of the Greengrass group.

bulkDeploymentResult_deploymentStatus :: Lens' BulkDeploymentResult (Maybe Text)

The current status of the group deployment: ''InProgress'', ''Building'', ''Success'', or ''Failure''.

bulkDeploymentResult_errorMessage :: Lens' BulkDeploymentResult (Maybe Text)

The error message for a failed deployment