Amazonka.Greengrass.Types.FunctionConfiguration

Description

Documentation

data FunctionConfiguration

The configuration of the Lambda function.

See: newFunctionConfiguration smart constructor.

Constructors

FunctionConfiguration'

Fields

memorySize :: Maybe Int

The memory size, in KB, which the function requires. This setting is not applicable and should be cleared when you run the Lambda function without containerization.

execArgs :: Maybe Text

The execution arguments.

environment :: Maybe FunctionConfigurationEnvironment

The environment configuration of the function.

executable :: Maybe Text

The name of the function executable.

pinned :: Maybe Bool

True if the function is pinned. Pinned means the function is long-lived and starts when the core starts.

encodingType :: Maybe EncodingType

The expected encoding type of the input payload for the function. The default is ''json''.

timeout :: Maybe Int

The allowed function execution time, after which Lambda should terminate the function. This timeout still applies to pinned Lambda functions for each request.

newFunctionConfiguration :: FunctionConfiguration

Create a value of FunctionConfiguration with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:memorySize:FunctionConfiguration', functionConfiguration_memorySize - The memory size, in KB, which the function requires. This setting is not applicable and should be cleared when you run the Lambda function without containerization.

$sel:execArgs:FunctionConfiguration', functionConfiguration_execArgs - The execution arguments.

$sel:environment:FunctionConfiguration', functionConfiguration_environment - The environment configuration of the function.

$sel:executable:FunctionConfiguration', functionConfiguration_executable - The name of the function executable.

$sel:pinned:FunctionConfiguration', functionConfiguration_pinned - True if the function is pinned. Pinned means the function is long-lived and starts when the core starts.

$sel:encodingType:FunctionConfiguration', functionConfiguration_encodingType - The expected encoding type of the input payload for the function. The default is ''json''.

$sel:timeout:FunctionConfiguration', functionConfiguration_timeout - The allowed function execution time, after which Lambda should terminate the function. This timeout still applies to pinned Lambda functions for each request.

functionConfiguration_memorySize :: Lens' FunctionConfiguration (Maybe Int)

The memory size, in KB, which the function requires. This setting is not applicable and should be cleared when you run the Lambda function without containerization.

functionConfiguration_execArgs :: Lens' FunctionConfiguration (Maybe Text)

The execution arguments.

functionConfiguration_environment :: Lens' FunctionConfiguration (Maybe FunctionConfigurationEnvironment)

The environment configuration of the function.

functionConfiguration_executable :: Lens' FunctionConfiguration (Maybe Text)

The name of the function executable.

functionConfiguration_pinned :: Lens' FunctionConfiguration (Maybe Bool)

True if the function is pinned. Pinned means the function is long-lived and starts when the core starts.

functionConfiguration_encodingType :: Lens' FunctionConfiguration (Maybe EncodingType)

The expected encoding type of the input payload for the function. The default is ''json''.

functionConfiguration_timeout :: Lens' FunctionConfiguration (Maybe Int)

The allowed function execution time, after which Lambda should terminate the function. This timeout still applies to pinned Lambda functions for each request.