Amazonka.Greengrass.Types.GroupInformation

Description

Documentation

data GroupInformation

Information about a group.

See: newGroupInformation smart constructor.

Constructors

GroupInformation'

Fields

latestVersionArn :: Maybe Text

The ARN of the latest version associated with the group.

arn :: Maybe Text

The ARN of the group.

name :: Maybe Text

The name of the group.

creationTimestamp :: Maybe Text

The time, in milliseconds since the epoch, when the group was created.

id :: Maybe Text

The ID of the group.

latestVersion :: Maybe Text

The ID of the latest version associated with the group.

lastUpdatedTimestamp :: Maybe Text

The time, in milliseconds since the epoch, when the group was last updated.

newGroupInformation :: GroupInformation

Create a value of GroupInformation with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:latestVersionArn:GroupInformation', groupInformation_latestVersionArn - The ARN of the latest version associated with the group.

$sel:arn:GroupInformation', groupInformation_arn - The ARN of the group.

$sel:name:GroupInformation', groupInformation_name - The name of the group.

$sel:creationTimestamp:GroupInformation', groupInformation_creationTimestamp - The time, in milliseconds since the epoch, when the group was created.

$sel:id:GroupInformation', groupInformation_id - The ID of the group.

$sel:latestVersion:GroupInformation', groupInformation_latestVersion - The ID of the latest version associated with the group.

$sel:lastUpdatedTimestamp:GroupInformation', groupInformation_lastUpdatedTimestamp - The time, in milliseconds since the epoch, when the group was last updated.

groupInformation_latestVersionArn :: Lens' GroupInformation (Maybe Text)

The ARN of the latest version associated with the group.

groupInformation_arn :: Lens' GroupInformation (Maybe Text)

The ARN of the group.

groupInformation_name :: Lens' GroupInformation (Maybe Text)

The name of the group.

groupInformation_creationTimestamp :: Lens' GroupInformation (Maybe Text)

The time, in milliseconds since the epoch, when the group was created.

groupInformation_id :: Lens' GroupInformation (Maybe Text)

The ID of the group.

groupInformation_latestVersion :: Lens' GroupInformation (Maybe Text)

The ID of the latest version associated with the group.

groupInformation_lastUpdatedTimestamp :: Lens' GroupInformation (Maybe Text)

The time, in milliseconds since the epoch, when the group was last updated.