Amazonka.GreengrassV2.BatchDisassociateClientDeviceFromCoreDevice

Description

Disassociate a list of client devices from a core device. After you disassociate a client device from a core device, the client device won't be able to use cloud discovery to retrieve the core device's connectivity information and certificates.

Creating a Request

data BatchDisassociateClientDeviceFromCoreDevice

See: newBatchDisassociateClientDeviceFromCoreDevice smart constructor.

Constructors

BatchDisassociateClientDeviceFromCoreDevice'

Fields

entries :: Maybe (NonEmpty DisassociateClientDeviceFromCoreDeviceEntry)

The list of client devices to disassociate.

coreDeviceThingName :: Text

The name of the core device. This is also the name of the IoT thing.

newBatchDisassociateClientDeviceFromCoreDevice

Arguments

:: Text
$sel:coreDeviceThingName:BatchDisassociateClientDeviceFromCoreDevice'

-> BatchDisassociateClientDeviceFromCoreDevice

Create a value of BatchDisassociateClientDeviceFromCoreDevice with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:entries:BatchDisassociateClientDeviceFromCoreDevice', batchDisassociateClientDeviceFromCoreDevice_entries - The list of client devices to disassociate.

$sel:coreDeviceThingName:BatchDisassociateClientDeviceFromCoreDevice', batchDisassociateClientDeviceFromCoreDevice_coreDeviceThingName - The name of the core device. This is also the name of the IoT thing.

Request Lenses

batchDisassociateClientDeviceFromCoreDevice_entries :: Lens' BatchDisassociateClientDeviceFromCoreDevice (Maybe (NonEmpty DisassociateClientDeviceFromCoreDeviceEntry))

The list of client devices to disassociate.

batchDisassociateClientDeviceFromCoreDevice_coreDeviceThingName :: Lens' BatchDisassociateClientDeviceFromCoreDevice Text

The name of the core device. This is also the name of the IoT thing.

Destructuring the Response

data BatchDisassociateClientDeviceFromCoreDeviceResponse

See: newBatchDisassociateClientDeviceFromCoreDeviceResponse smart constructor.

Constructors

BatchDisassociateClientDeviceFromCoreDeviceResponse'

Fields

errorEntries :: Maybe (NonEmpty DisassociateClientDeviceFromCoreDeviceErrorEntry)

The list of errors (if any) for the entries in the request. Each error entry contains the name of the IoT thing that failed to disassociate.

httpStatus :: Int

The response's http status code.

newBatchDisassociateClientDeviceFromCoreDeviceResponse

Arguments

:: Int
$sel:httpStatus:BatchDisassociateClientDeviceFromCoreDeviceResponse'

-> BatchDisassociateClientDeviceFromCoreDeviceResponse

Create a value of BatchDisassociateClientDeviceFromCoreDeviceResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:errorEntries:BatchDisassociateClientDeviceFromCoreDeviceResponse', batchDisassociateClientDeviceFromCoreDeviceResponse_errorEntries - The list of errors (if any) for the entries in the request. Each error entry contains the name of the IoT thing that failed to disassociate.

$sel:httpStatus:BatchDisassociateClientDeviceFromCoreDeviceResponse', batchDisassociateClientDeviceFromCoreDeviceResponse_httpStatus - The response's http status code.

Response Lenses

batchDisassociateClientDeviceFromCoreDeviceResponse_errorEntries :: Lens' BatchDisassociateClientDeviceFromCoreDeviceResponse (Maybe (NonEmpty DisassociateClientDeviceFromCoreDeviceErrorEntry))

The list of errors (if any) for the entries in the request. Each error entry contains the name of the IoT thing that failed to disassociate.

batchDisassociateClientDeviceFromCoreDeviceResponse_httpStatus :: Lens' BatchDisassociateClientDeviceFromCoreDeviceResponse Int

The response's http status code.