Amazonka.GreengrassV2.CreateDeployment

Description

Creates a continuous deployment for a target, which is a Greengrass core device or group of core devices. When you add a new core device to a group of core devices that has a deployment, IoT Greengrass deploys that group's deployment to the new device.

You can define one deployment for each target. When you create a new deployment for a target that has an existing deployment, you replace the previous deployment. IoT Greengrass applies the new deployment to the target devices.

Every deployment has a revision number that indicates how many deployment revisions you define for a target. Use this operation to create a new revision of an existing deployment. This operation returns the revision number of the new deployment when you create it.

For more information, see the Create deployments in the IoT Greengrass V2 Developer Guide.

Creating a Request

data CreateDeployment

See: newCreateDeployment smart constructor.

Constructors

CreateDeployment'

Fields

components :: Maybe (HashMap Text ComponentDeploymentSpecification)

The components to deploy. This is a dictionary, where each key is the name of a component, and each key's value is the version and configuration to deploy for that component.

clientToken :: Maybe Text

A unique, case-sensitive identifier that you can provide to ensure that the request is idempotent. Idempotency means that the request is successfully processed only once, even if you send the request multiple times. When a request succeeds, and you specify the same client token for subsequent successful requests, the IoT Greengrass V2 service returns the successful response that it caches from the previous request. IoT Greengrass V2 caches successful responses for idempotent requests for up to 8 hours.

deploymentPolicies :: Maybe DeploymentPolicies

The deployment policies for the deployment. These policies define how the deployment updates components and handles failure.

iotJobConfiguration :: Maybe DeploymentIoTJobConfiguration

The job configuration for the deployment configuration. The job configuration specifies the rollout, timeout, and stop configurations for the deployment configuration.

deploymentName :: Maybe Text

The name of the deployment.

tags :: Maybe (HashMap Text Text)

A list of key-value pairs that contain metadata for the resource. For more information, see Tag your resources in the IoT Greengrass V2 Developer Guide.

targetArn :: Text

The ARN of the target IoT thing or thing group.

newCreateDeployment

Arguments

:: Text
$sel:targetArn:CreateDeployment'

-> CreateDeployment

Create a value of CreateDeployment with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:components:CreateDeployment', createDeployment_components - The components to deploy. This is a dictionary, where each key is the name of a component, and each key's value is the version and configuration to deploy for that component.

$sel:clientToken:CreateDeployment', createDeployment_clientToken - A unique, case-sensitive identifier that you can provide to ensure that the request is idempotent. Idempotency means that the request is successfully processed only once, even if you send the request multiple times. When a request succeeds, and you specify the same client token for subsequent successful requests, the IoT Greengrass V2 service returns the successful response that it caches from the previous request. IoT Greengrass V2 caches successful responses for idempotent requests for up to 8 hours.

$sel:deploymentPolicies:CreateDeployment', createDeployment_deploymentPolicies - The deployment policies for the deployment. These policies define how the deployment updates components and handles failure.

$sel:iotJobConfiguration:CreateDeployment', createDeployment_iotJobConfiguration - The job configuration for the deployment configuration. The job configuration specifies the rollout, timeout, and stop configurations for the deployment configuration.

$sel:deploymentName:CreateDeployment', createDeployment_deploymentName - The name of the deployment.

$sel:tags:CreateDeployment', createDeployment_tags - A list of key-value pairs that contain metadata for the resource. For more information, see Tag your resources in the IoT Greengrass V2 Developer Guide.

$sel:targetArn:CreateDeployment', createDeployment_targetArn - The ARN of the target IoT thing or thing group.

Request Lenses

createDeployment_components :: Lens' CreateDeployment (Maybe (HashMap Text ComponentDeploymentSpecification))

The components to deploy. This is a dictionary, where each key is the name of a component, and each key's value is the version and configuration to deploy for that component.

createDeployment_clientToken :: Lens' CreateDeployment (Maybe Text)

A unique, case-sensitive identifier that you can provide to ensure that the request is idempotent. Idempotency means that the request is successfully processed only once, even if you send the request multiple times. When a request succeeds, and you specify the same client token for subsequent successful requests, the IoT Greengrass V2 service returns the successful response that it caches from the previous request. IoT Greengrass V2 caches successful responses for idempotent requests for up to 8 hours.

createDeployment_deploymentPolicies :: Lens' CreateDeployment (Maybe DeploymentPolicies)

The deployment policies for the deployment. These policies define how the deployment updates components and handles failure.

createDeployment_iotJobConfiguration :: Lens' CreateDeployment (Maybe DeploymentIoTJobConfiguration)

The job configuration for the deployment configuration. The job configuration specifies the rollout, timeout, and stop configurations for the deployment configuration.

createDeployment_deploymentName :: Lens' CreateDeployment (Maybe Text)

The name of the deployment.

createDeployment_tags :: Lens' CreateDeployment (Maybe (HashMap Text Text))

A list of key-value pairs that contain metadata for the resource. For more information, see Tag your resources in the IoT Greengrass V2 Developer Guide.

createDeployment_targetArn :: Lens' CreateDeployment Text

The ARN of the target IoT thing or thing group.

Destructuring the Response

data CreateDeploymentResponse

See: newCreateDeploymentResponse smart constructor.

Constructors

CreateDeploymentResponse'

Fields

deploymentId :: Maybe Text

The ID of the deployment.

iotJobId :: Maybe Text

The ID of the IoT job that applies the deployment to target devices.

iotJobArn :: Maybe Text

The ARN of the IoT job that applies the deployment to target devices.

httpStatus :: Int

The response's http status code.

newCreateDeploymentResponse

Arguments

:: Int
$sel:httpStatus:CreateDeploymentResponse'

-> CreateDeploymentResponse

Create a value of CreateDeploymentResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:deploymentId:CreateDeploymentResponse', createDeploymentResponse_deploymentId - The ID of the deployment.

$sel:iotJobId:CreateDeploymentResponse', createDeploymentResponse_iotJobId - The ID of the IoT job that applies the deployment to target devices.

$sel:iotJobArn:CreateDeploymentResponse', createDeploymentResponse_iotJobArn - The ARN of the IoT job that applies the deployment to target devices.

$sel:httpStatus:CreateDeploymentResponse', createDeploymentResponse_httpStatus - The response's http status code.

Response Lenses

createDeploymentResponse_deploymentId :: Lens' CreateDeploymentResponse (Maybe Text)

The ID of the deployment.

createDeploymentResponse_iotJobId :: Lens' CreateDeploymentResponse (Maybe Text)

The ID of the IoT job that applies the deployment to target devices.

createDeploymentResponse_iotJobArn :: Lens' CreateDeploymentResponse (Maybe Text)

The ARN of the IoT job that applies the deployment to target devices.

createDeploymentResponse_httpStatus :: Lens' CreateDeploymentResponse Int

The response's http status code.