Amazonka.GreengrassV2.DescribeComponent

Description

Retrieves metadata for a version of a component.

Creating a Request

data DescribeComponent

See: newDescribeComponent smart constructor.

Constructors

DescribeComponent'

Fields

arn :: Text

The ARN of the component version.

newDescribeComponent

Arguments

:: Text
$sel:arn:DescribeComponent'

-> DescribeComponent

Create a value of DescribeComponent with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:arn:DescribeComponent', describeComponent_arn - The ARN of the component version.

Request Lenses

describeComponent_arn :: Lens' DescribeComponent Text

The ARN of the component version.

Destructuring the Response

data DescribeComponentResponse

See: newDescribeComponentResponse smart constructor.

Constructors

DescribeComponentResponse'

Fields

platforms :: Maybe [ComponentPlatform]

The platforms that the component version supports.

status :: Maybe CloudComponentStatus

The status of the component version in IoT Greengrass V2. This status is different from the status of the component on a core device.

arn :: Maybe Text

The ARN of the component version.

componentVersion :: Maybe Text

The version of the component.

creationTimestamp :: Maybe POSIX

The time at which the component was created, expressed in ISO 8601 format.

componentName :: Maybe Text

The name of the component.

publisher :: Maybe Text

The publisher of the component version.

description :: Maybe Text

The description of the component version.

tags :: Maybe (HashMap Text Text)

A list of key-value pairs that contain metadata for the resource. For more information, see Tag your resources in the IoT Greengrass V2 Developer Guide.

httpStatus :: Int

The response's http status code.

newDescribeComponentResponse

Arguments

:: Int
$sel:httpStatus:DescribeComponentResponse'

-> DescribeComponentResponse

Create a value of DescribeComponentResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:platforms:DescribeComponentResponse', describeComponentResponse_platforms - The platforms that the component version supports.

$sel:status:DescribeComponentResponse', describeComponentResponse_status - The status of the component version in IoT Greengrass V2. This status is different from the status of the component on a core device.

$sel:arn:DescribeComponent', describeComponentResponse_arn - The ARN of the component version.

$sel:componentVersion:DescribeComponentResponse', describeComponentResponse_componentVersion - The version of the component.

$sel:creationTimestamp:DescribeComponentResponse', describeComponentResponse_creationTimestamp - The time at which the component was created, expressed in ISO 8601 format.

$sel:componentName:DescribeComponentResponse', describeComponentResponse_componentName - The name of the component.

$sel:publisher:DescribeComponentResponse', describeComponentResponse_publisher - The publisher of the component version.

$sel:description:DescribeComponentResponse', describeComponentResponse_description - The description of the component version.

$sel:tags:DescribeComponentResponse', describeComponentResponse_tags - A list of key-value pairs that contain metadata for the resource. For more information, see Tag your resources in the IoT Greengrass V2 Developer Guide.

$sel:httpStatus:DescribeComponentResponse', describeComponentResponse_httpStatus - The response's http status code.

Response Lenses

describeComponentResponse_platforms :: Lens' DescribeComponentResponse (Maybe [ComponentPlatform])

The platforms that the component version supports.

describeComponentResponse_status :: Lens' DescribeComponentResponse (Maybe CloudComponentStatus)

The status of the component version in IoT Greengrass V2. This status is different from the status of the component on a core device.

describeComponentResponse_arn :: Lens' DescribeComponentResponse (Maybe Text)

The ARN of the component version.

describeComponentResponse_componentVersion :: Lens' DescribeComponentResponse (Maybe Text)

The version of the component.

describeComponentResponse_creationTimestamp :: Lens' DescribeComponentResponse (Maybe UTCTime)

The time at which the component was created, expressed in ISO 8601 format.

describeComponentResponse_componentName :: Lens' DescribeComponentResponse (Maybe Text)

The name of the component.

describeComponentResponse_publisher :: Lens' DescribeComponentResponse (Maybe Text)

The publisher of the component version.

describeComponentResponse_description :: Lens' DescribeComponentResponse (Maybe Text)

The description of the component version.

describeComponentResponse_tags :: Lens' DescribeComponentResponse (Maybe (HashMap Text Text))

A list of key-value pairs that contain metadata for the resource. For more information, see Tag your resources in the IoT Greengrass V2 Developer Guide.

describeComponentResponse_httpStatus :: Lens' DescribeComponentResponse Int

The response's http status code.