Amazonka.GreengrassV2.GetCoreDevice

Description

Retrieves metadata for a Greengrass core device.

Creating a Request

data GetCoreDevice

See: newGetCoreDevice smart constructor.

Constructors

GetCoreDevice'

Fields

coreDeviceThingName :: Text

The name of the core device. This is also the name of the IoT thing.

newGetCoreDevice

Arguments

:: Text
$sel:coreDeviceThingName:GetCoreDevice'

-> GetCoreDevice

Create a value of GetCoreDevice with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:coreDeviceThingName:GetCoreDevice', getCoreDevice_coreDeviceThingName - The name of the core device. This is also the name of the IoT thing.

Request Lenses

getCoreDevice_coreDeviceThingName :: Lens' GetCoreDevice Text

The name of the core device. This is also the name of the IoT thing.

Destructuring the Response

data GetCoreDeviceResponse

See: newGetCoreDeviceResponse smart constructor.

Constructors

GetCoreDeviceResponse'

Fields

status :: Maybe CoreDeviceStatus

The status of the core device. The core device status can be:

HEALTHY – The IoT Greengrass Core software and all components run on the core device without issue.

UNHEALTHY – The IoT Greengrass Core software or a component is in a failed state on the core device.

platform :: Maybe Text

The operating system platform that the core device runs.

architecture :: Maybe Text

The computer architecture of the core device.

coreDeviceThingName :: Maybe Text

The name of the core device. This is also the name of the IoT thing.

tags :: Maybe (HashMap Text Text)

A list of key-value pairs that contain metadata for the resource. For more information, see Tag your resources in the IoT Greengrass V2 Developer Guide.

coreVersion :: Maybe Text

The version of the IoT Greengrass Core software that the core device runs. This version is equivalent to the version of the Greengrass nucleus component that runs on the core device. For more information, see the Greengrass nucleus component in the IoT Greengrass V2 Developer Guide.

lastStatusUpdateTimestamp :: Maybe POSIX

The time at which the core device's status last updated, expressed in ISO 8601 format.

httpStatus :: Int

The response's http status code.

newGetCoreDeviceResponse

Arguments

:: Int
$sel:httpStatus:GetCoreDeviceResponse'

-> GetCoreDeviceResponse

Create a value of GetCoreDeviceResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:status:GetCoreDeviceResponse', getCoreDeviceResponse_status - The status of the core device. The core device status can be:

HEALTHY – The IoT Greengrass Core software and all components run on the core device without issue.

UNHEALTHY – The IoT Greengrass Core software or a component is in a failed state on the core device.

$sel:platform:GetCoreDeviceResponse', getCoreDeviceResponse_platform - The operating system platform that the core device runs.

$sel:architecture:GetCoreDeviceResponse', getCoreDeviceResponse_architecture - The computer architecture of the core device.

$sel:coreDeviceThingName:GetCoreDevice', getCoreDeviceResponse_coreDeviceThingName - The name of the core device. This is also the name of the IoT thing.

$sel:tags:GetCoreDeviceResponse', getCoreDeviceResponse_tags - A list of key-value pairs that contain metadata for the resource. For more information, see Tag your resources in the IoT Greengrass V2 Developer Guide.

$sel:coreVersion:GetCoreDeviceResponse', getCoreDeviceResponse_coreVersion - The version of the IoT Greengrass Core software that the core device runs. This version is equivalent to the version of the Greengrass nucleus component that runs on the core device. For more information, see the Greengrass nucleus component in the IoT Greengrass V2 Developer Guide.

$sel:lastStatusUpdateTimestamp:GetCoreDeviceResponse', getCoreDeviceResponse_lastStatusUpdateTimestamp - The time at which the core device's status last updated, expressed in ISO 8601 format.

$sel:httpStatus:GetCoreDeviceResponse', getCoreDeviceResponse_httpStatus - The response's http status code.

Response Lenses

getCoreDeviceResponse_status :: Lens' GetCoreDeviceResponse (Maybe CoreDeviceStatus)

The status of the core device. The core device status can be:

HEALTHY – The IoT Greengrass Core software and all components run on the core device without issue.

UNHEALTHY – The IoT Greengrass Core software or a component is in a failed state on the core device.

getCoreDeviceResponse_platform :: Lens' GetCoreDeviceResponse (Maybe Text)

The operating system platform that the core device runs.

getCoreDeviceResponse_architecture :: Lens' GetCoreDeviceResponse (Maybe Text)

The computer architecture of the core device.

getCoreDeviceResponse_coreDeviceThingName :: Lens' GetCoreDeviceResponse (Maybe Text)

The name of the core device. This is also the name of the IoT thing.

getCoreDeviceResponse_tags :: Lens' GetCoreDeviceResponse (Maybe (HashMap Text Text))

A list of key-value pairs that contain metadata for the resource. For more information, see Tag your resources in the IoT Greengrass V2 Developer Guide.

getCoreDeviceResponse_coreVersion :: Lens' GetCoreDeviceResponse (Maybe Text)

The version of the IoT Greengrass Core software that the core device runs. This version is equivalent to the version of the Greengrass nucleus component that runs on the core device. For more information, see the Greengrass nucleus component in the IoT Greengrass V2 Developer Guide.

getCoreDeviceResponse_lastStatusUpdateTimestamp :: Lens' GetCoreDeviceResponse (Maybe UTCTime)

The time at which the core device's status last updated, expressed in ISO 8601 format.

getCoreDeviceResponse_httpStatus :: Lens' GetCoreDeviceResponse Int

The response's http status code.