Amazonka.GreengrassV2.ListCoreDevices

Description

Retrieves a paginated list of Greengrass core devices.

This operation returns paginated results.

Creating a Request

data ListCoreDevices

See: newListCoreDevices smart constructor.

Constructors

ListCoreDevices'

Fields

status :: Maybe CoreDeviceStatus

The core device status by which to filter. If you specify this parameter, the list includes only core devices that have this status. Choose one of the following options:

HEALTHY – The IoT Greengrass Core software and all components run on the core device without issue.

UNHEALTHY – The IoT Greengrass Core software or a component is in a failed state on the core device.

thingGroupArn :: Maybe Text

The ARN of the IoT thing group by which to filter. If you specify this parameter, the list includes only core devices that are members of this thing group.

nextToken :: Maybe Text

The token to be used for the next set of paginated results.

maxResults :: Maybe Natural

The maximum number of results to be returned per paginated request.

newListCoreDevices :: ListCoreDevices

Create a value of ListCoreDevices with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:status:ListCoreDevices', listCoreDevices_status - The core device status by which to filter. If you specify this parameter, the list includes only core devices that have this status. Choose one of the following options:

HEALTHY – The IoT Greengrass Core software and all components run on the core device without issue.

UNHEALTHY – The IoT Greengrass Core software or a component is in a failed state on the core device.

$sel:thingGroupArn:ListCoreDevices', listCoreDevices_thingGroupArn - The ARN of the IoT thing group by which to filter. If you specify this parameter, the list includes only core devices that are members of this thing group.

$sel:nextToken:ListCoreDevices', listCoreDevices_nextToken - The token to be used for the next set of paginated results.

$sel:maxResults:ListCoreDevices', listCoreDevices_maxResults - The maximum number of results to be returned per paginated request.

Request Lenses

listCoreDevices_status :: Lens' ListCoreDevices (Maybe CoreDeviceStatus)

The core device status by which to filter. If you specify this parameter, the list includes only core devices that have this status. Choose one of the following options:

HEALTHY – The IoT Greengrass Core software and all components run on the core device without issue.

UNHEALTHY – The IoT Greengrass Core software or a component is in a failed state on the core device.

listCoreDevices_thingGroupArn :: Lens' ListCoreDevices (Maybe Text)

The ARN of the IoT thing group by which to filter. If you specify this parameter, the list includes only core devices that are members of this thing group.

listCoreDevices_nextToken :: Lens' ListCoreDevices (Maybe Text)

The token to be used for the next set of paginated results.

listCoreDevices_maxResults :: Lens' ListCoreDevices (Maybe Natural)

The maximum number of results to be returned per paginated request.

Destructuring the Response

data ListCoreDevicesResponse

See: newListCoreDevicesResponse smart constructor.

Constructors

ListCoreDevicesResponse'

Fields

nextToken :: Maybe Text

The token for the next set of results, or null if there are no additional results.

coreDevices :: Maybe [CoreDevice]

A list that summarizes each core device.

httpStatus :: Int

The response's http status code.

newListCoreDevicesResponse

Arguments

:: Int
$sel:httpStatus:ListCoreDevicesResponse'

-> ListCoreDevicesResponse

Create a value of ListCoreDevicesResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:nextToken:ListCoreDevices', listCoreDevicesResponse_nextToken - The token for the next set of results, or null if there are no additional results.

$sel:coreDevices:ListCoreDevicesResponse', listCoreDevicesResponse_coreDevices - A list that summarizes each core device.

$sel:httpStatus:ListCoreDevicesResponse', listCoreDevicesResponse_httpStatus - The response's http status code.

Response Lenses

listCoreDevicesResponse_nextToken :: Lens' ListCoreDevicesResponse (Maybe Text)

The token for the next set of results, or null if there are no additional results.

listCoreDevicesResponse_coreDevices :: Lens' ListCoreDevicesResponse (Maybe [CoreDevice])

A list that summarizes each core device.

listCoreDevicesResponse_httpStatus :: Lens' ListCoreDevicesResponse Int

The response's http status code.