Amazonka.GreengrassV2.ListInstalledComponents

Description

Retrieves a paginated list of the components that a Greengrass core device runs.

This operation returns paginated results.

Creating a Request

data ListInstalledComponents

See: newListInstalledComponents smart constructor.

Constructors

ListInstalledComponents'

Fields

nextToken :: Maybe Text

The token to be used for the next set of paginated results.

maxResults :: Maybe Natural

The maximum number of results to be returned per paginated request.

coreDeviceThingName :: Text

The name of the core device. This is also the name of the IoT thing.

newListInstalledComponents

Arguments

:: Text
$sel:coreDeviceThingName:ListInstalledComponents'

-> ListInstalledComponents

Create a value of ListInstalledComponents with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:nextToken:ListInstalledComponents', listInstalledComponents_nextToken - The token to be used for the next set of paginated results.

$sel:maxResults:ListInstalledComponents', listInstalledComponents_maxResults - The maximum number of results to be returned per paginated request.

$sel:coreDeviceThingName:ListInstalledComponents', listInstalledComponents_coreDeviceThingName - The name of the core device. This is also the name of the IoT thing.

Request Lenses

listInstalledComponents_nextToken :: Lens' ListInstalledComponents (Maybe Text)

The token to be used for the next set of paginated results.

listInstalledComponents_maxResults :: Lens' ListInstalledComponents (Maybe Natural)

The maximum number of results to be returned per paginated request.

listInstalledComponents_coreDeviceThingName :: Lens' ListInstalledComponents Text

The name of the core device. This is also the name of the IoT thing.

Destructuring the Response

data ListInstalledComponentsResponse

See: newListInstalledComponentsResponse smart constructor.

Constructors

ListInstalledComponentsResponse'

Fields

installedComponents :: Maybe [InstalledComponent]

A list that summarizes each component on the core device.

nextToken :: Maybe Text

The token for the next set of results, or null if there are no additional results.

httpStatus :: Int

The response's http status code.

newListInstalledComponentsResponse

Arguments

:: Int
$sel:httpStatus:ListInstalledComponentsResponse'

-> ListInstalledComponentsResponse

Create a value of ListInstalledComponentsResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:installedComponents:ListInstalledComponentsResponse', listInstalledComponentsResponse_installedComponents - A list that summarizes each component on the core device.

$sel:nextToken:ListInstalledComponents', listInstalledComponentsResponse_nextToken - The token for the next set of results, or null if there are no additional results.

$sel:httpStatus:ListInstalledComponentsResponse', listInstalledComponentsResponse_httpStatus - The response's http status code.

Response Lenses

listInstalledComponentsResponse_installedComponents :: Lens' ListInstalledComponentsResponse (Maybe [InstalledComponent])

A list that summarizes each component on the core device.

listInstalledComponentsResponse_nextToken :: Lens' ListInstalledComponentsResponse (Maybe Text)

The token for the next set of results, or null if there are no additional results.

listInstalledComponentsResponse_httpStatus :: Lens' ListInstalledComponentsResponse Int

The response's http status code.