Amazonka.GreengrassV2.ResolveComponentCandidates

Description

Retrieves a list of components that meet the component, version, and platform requirements of a deployment. Greengrass core devices call this operation when they receive a deployment to identify the components to install.

This operation identifies components that meet all dependency requirements for a deployment. If the requirements conflict, then this operation returns an error and the deployment fails. For example, this occurs if component A requires version >2.0.0 and component B requires version <2.0.0 of a component dependency.

When you specify the component candidates to resolve, IoT Greengrass compares each component's digest from the core device with the component's digest in the Amazon Web Services Cloud. If the digests don't match, then IoT Greengrass specifies to use the version from the Amazon Web Services Cloud.

To use this operation, you must use the data plane API endpoint and authenticate with an IoT device certificate. For more information, see IoT Greengrass endpoints and quotas.

Creating a Request

data ResolveComponentCandidates

See: newResolveComponentCandidates smart constructor.

Constructors

ResolveComponentCandidates'

Fields

platform :: ComponentPlatform

The platform to use to resolve compatible components.

componentCandidates :: [ComponentCandidate]

The list of components to resolve.

newResolveComponentCandidates

Arguments

:: ComponentPlatform
$sel:platform:ResolveComponentCandidates'

-> ResolveComponentCandidates

Create a value of ResolveComponentCandidates with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:platform:ResolveComponentCandidates', resolveComponentCandidates_platform - The platform to use to resolve compatible components.

$sel:componentCandidates:ResolveComponentCandidates', resolveComponentCandidates_componentCandidates - The list of components to resolve.

Request Lenses

resolveComponentCandidates_platform :: Lens' ResolveComponentCandidates ComponentPlatform

The platform to use to resolve compatible components.

resolveComponentCandidates_componentCandidates :: Lens' ResolveComponentCandidates [ComponentCandidate]

The list of components to resolve.

Destructuring the Response

data ResolveComponentCandidatesResponse

See: newResolveComponentCandidatesResponse smart constructor.

Constructors

ResolveComponentCandidatesResponse'

Fields

resolvedComponentVersions :: Maybe [ResolvedComponentVersion]

A list of components that meet the requirements that you specify in the request. This list includes each component's recipe that you can use to install the component.

httpStatus :: Int

The response's http status code.

newResolveComponentCandidatesResponse

Arguments

:: Int
$sel:httpStatus:ResolveComponentCandidatesResponse'

-> ResolveComponentCandidatesResponse

Create a value of ResolveComponentCandidatesResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:resolvedComponentVersions:ResolveComponentCandidatesResponse', resolveComponentCandidatesResponse_resolvedComponentVersions - A list of components that meet the requirements that you specify in the request. This list includes each component's recipe that you can use to install the component.

$sel:httpStatus:ResolveComponentCandidatesResponse', resolveComponentCandidatesResponse_httpStatus - The response's http status code.

Response Lenses

resolveComponentCandidatesResponse_resolvedComponentVersions :: Lens' ResolveComponentCandidatesResponse (Maybe [ResolvedComponentVersion])

A list of components that meet the requirements that you specify in the request. This list includes each component's recipe that you can use to install the component.

resolveComponentCandidatesResponse_httpStatus :: Lens' ResolveComponentCandidatesResponse Int

The response's http status code.