Amazonka.GreengrassV2.Types.AssociateClientDeviceWithCoreDeviceErrorEntry

Description

Documentation

data AssociateClientDeviceWithCoreDeviceErrorEntry

Contains an error that occurs from a request to associate a client device with a core device. The BatchAssociateClientDeviceWithCoreDevice operation returns a list of these errors.

See: newAssociateClientDeviceWithCoreDeviceErrorEntry smart constructor.

Constructors

AssociateClientDeviceWithCoreDeviceErrorEntry'

Fields

code :: Maybe Text

The error code for the request.

message :: Maybe Text

A message that provides additional information about the error.

thingName :: Maybe Text

The name of the IoT thing whose associate request failed.

newAssociateClientDeviceWithCoreDeviceErrorEntry :: AssociateClientDeviceWithCoreDeviceErrorEntry

Create a value of AssociateClientDeviceWithCoreDeviceErrorEntry with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:code:AssociateClientDeviceWithCoreDeviceErrorEntry', associateClientDeviceWithCoreDeviceErrorEntry_code - The error code for the request.

$sel:message:AssociateClientDeviceWithCoreDeviceErrorEntry', associateClientDeviceWithCoreDeviceErrorEntry_message - A message that provides additional information about the error.

$sel:thingName:AssociateClientDeviceWithCoreDeviceErrorEntry', associateClientDeviceWithCoreDeviceErrorEntry_thingName - The name of the IoT thing whose associate request failed.

associateClientDeviceWithCoreDeviceErrorEntry_code :: Lens' AssociateClientDeviceWithCoreDeviceErrorEntry (Maybe Text)

The error code for the request.

associateClientDeviceWithCoreDeviceErrorEntry_message :: Lens' AssociateClientDeviceWithCoreDeviceErrorEntry (Maybe Text)

A message that provides additional information about the error.

associateClientDeviceWithCoreDeviceErrorEntry_thingName :: Lens' AssociateClientDeviceWithCoreDeviceErrorEntry (Maybe Text)

The name of the IoT thing whose associate request failed.