Amazonka.GreengrassV2.Types.ComponentCandidate

Description

Documentation

data ComponentCandidate

Contains information about a component that is a candidate to deploy to a Greengrass core device.

See: newComponentCandidate smart constructor.

Constructors

ComponentCandidate'

Fields

componentVersion :: Maybe Text

The version of the component.

versionRequirements :: Maybe (HashMap Text Text)

The version requirements for the component's dependencies. Greengrass core devices get the version requirements from component recipes.

IoT Greengrass V2 uses semantic version constraints. For more information, see Semantic Versioning.

componentName :: Maybe Text

The name of the component.

newComponentCandidate :: ComponentCandidate

Create a value of ComponentCandidate with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:componentVersion:ComponentCandidate', componentCandidate_componentVersion - The version of the component.

$sel:versionRequirements:ComponentCandidate', componentCandidate_versionRequirements - The version requirements for the component's dependencies. Greengrass core devices get the version requirements from component recipes.

IoT Greengrass V2 uses semantic version constraints. For more information, see Semantic Versioning.

$sel:componentName:ComponentCandidate', componentCandidate_componentName - The name of the component.

componentCandidate_componentVersion :: Lens' ComponentCandidate (Maybe Text)

The version of the component.

componentCandidate_versionRequirements :: Lens' ComponentCandidate (Maybe (HashMap Text Text))

The version requirements for the component's dependencies. Greengrass core devices get the version requirements from component recipes.

IoT Greengrass V2 uses semantic version constraints. For more information, see Semantic Versioning.

componentCandidate_componentName :: Lens' ComponentCandidate (Maybe Text)

The name of the component.