Amazonka.GreengrassV2.Types.CoreDevice

Description

Documentation

data CoreDevice

Contains information about a Greengrass core device, which is an IoT thing that runs the IoT Greengrass Core software.

See: newCoreDevice smart constructor.

Constructors

CoreDevice'

Fields

status :: Maybe CoreDeviceStatus

The status of the core device. Core devices can have the following statuses:

HEALTHY – The IoT Greengrass Core software and all components run on the core device without issue.

UNHEALTHY – The IoT Greengrass Core software or a component is in a failed state on the core device.

coreDeviceThingName :: Maybe Text

The name of the core device. This is also the name of the IoT thing.

lastStatusUpdateTimestamp :: Maybe POSIX

The time at which the core device's status last updated, expressed in ISO 8601 format.

newCoreDevice :: CoreDevice

Create a value of CoreDevice with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:status:CoreDevice', coreDevice_status - The status of the core device. Core devices can have the following statuses:

HEALTHY – The IoT Greengrass Core software and all components run on the core device without issue.

UNHEALTHY – The IoT Greengrass Core software or a component is in a failed state on the core device.

$sel:coreDeviceThingName:CoreDevice', coreDevice_coreDeviceThingName - The name of the core device. This is also the name of the IoT thing.

$sel:lastStatusUpdateTimestamp:CoreDevice', coreDevice_lastStatusUpdateTimestamp - The time at which the core device's status last updated, expressed in ISO 8601 format.

coreDevice_status :: Lens' CoreDevice (Maybe CoreDeviceStatus)

The status of the core device. Core devices can have the following statuses:

HEALTHY – The IoT Greengrass Core software and all components run on the core device without issue.

UNHEALTHY – The IoT Greengrass Core software or a component is in a failed state on the core device.

coreDevice_coreDeviceThingName :: Lens' CoreDevice (Maybe Text)

The name of the core device. This is also the name of the IoT thing.

coreDevice_lastStatusUpdateTimestamp :: Lens' CoreDevice (Maybe UTCTime)

The time at which the core device's status last updated, expressed in ISO 8601 format.