Amazonka.GreengrassV2.Types.InstalledComponent

Description

Documentation

data InstalledComponent

Contains information about a component on a Greengrass core device.

See: newInstalledComponent smart constructor.

Constructors

InstalledComponent'

Fields

isRoot :: Maybe Bool

Whether or not the component is a root component.

componentVersion :: Maybe Text

The version of the component.

componentName :: Maybe Text

The name of the component.

lifecycleStateDetails :: Maybe Text

The details about the lifecycle state of the component.

lifecycleState :: Maybe InstalledComponentLifecycleState

The lifecycle state of the component.

newInstalledComponent :: InstalledComponent

Create a value of InstalledComponent with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:isRoot:InstalledComponent', installedComponent_isRoot - Whether or not the component is a root component.

$sel:componentVersion:InstalledComponent', installedComponent_componentVersion - The version of the component.

$sel:componentName:InstalledComponent', installedComponent_componentName - The name of the component.

$sel:lifecycleStateDetails:InstalledComponent', installedComponent_lifecycleStateDetails - The details about the lifecycle state of the component.

$sel:lifecycleState:InstalledComponent', installedComponent_lifecycleState - The lifecycle state of the component.

installedComponent_isRoot :: Lens' InstalledComponent (Maybe Bool)

Whether or not the component is a root component.

installedComponent_componentVersion :: Lens' InstalledComponent (Maybe Text)

The version of the component.

installedComponent_componentName :: Lens' InstalledComponent (Maybe Text)

The name of the component.

installedComponent_lifecycleStateDetails :: Lens' InstalledComponent (Maybe Text)

The details about the lifecycle state of the component.

installedComponent_lifecycleState :: Lens' InstalledComponent (Maybe InstalledComponentLifecycleState)

The lifecycle state of the component.