Amazonka.GreengrassV2.Types.IoTJobExponentialRolloutRate

Description

Documentation

data IoTJobExponentialRolloutRate

Contains information about an exponential rollout rate for a configuration deployment job.

See: newIoTJobExponentialRolloutRate smart constructor.

Constructors

IoTJobExponentialRolloutRate'

Fields

baseRatePerMinute :: Natural

The minimum number of devices that receive a pending job notification, per minute, when the job starts. This parameter defines the initial rollout rate of the job.

incrementFactor :: Double

The exponential factor to increase the rollout rate for the job.

This parameter supports up to one digit after the decimal (for example, you can specify 1.5, but not 1.55).

rateIncreaseCriteria :: IoTJobRateIncreaseCriteria

The criteria to increase the rollout rate for the job.

newIoTJobExponentialRolloutRate

Arguments

:: Natural
$sel:baseRatePerMinute:IoTJobExponentialRolloutRate'

-> Double
$sel:incrementFactor:IoTJobExponentialRolloutRate'

-> IoTJobRateIncreaseCriteria
$sel:rateIncreaseCriteria:IoTJobExponentialRolloutRate'

-> IoTJobExponentialRolloutRate

Create a value of IoTJobExponentialRolloutRate with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:baseRatePerMinute:IoTJobExponentialRolloutRate', ioTJobExponentialRolloutRate_baseRatePerMinute - The minimum number of devices that receive a pending job notification, per minute, when the job starts. This parameter defines the initial rollout rate of the job.

$sel:incrementFactor:IoTJobExponentialRolloutRate', ioTJobExponentialRolloutRate_incrementFactor - The exponential factor to increase the rollout rate for the job.

This parameter supports up to one digit after the decimal (for example, you can specify 1.5, but not 1.55).

$sel:rateIncreaseCriteria:IoTJobExponentialRolloutRate', ioTJobExponentialRolloutRate_rateIncreaseCriteria - The criteria to increase the rollout rate for the job.

ioTJobExponentialRolloutRate_baseRatePerMinute :: Lens' IoTJobExponentialRolloutRate Natural

The minimum number of devices that receive a pending job notification, per minute, when the job starts. This parameter defines the initial rollout rate of the job.

ioTJobExponentialRolloutRate_incrementFactor :: Lens' IoTJobExponentialRolloutRate Double

The exponential factor to increase the rollout rate for the job.

This parameter supports up to one digit after the decimal (for example, you can specify 1.5, but not 1.55).

ioTJobExponentialRolloutRate_rateIncreaseCriteria :: Lens' IoTJobExponentialRolloutRate IoTJobRateIncreaseCriteria

The criteria to increase the rollout rate for the job.