Amazonka.GreengrassV2.Types.LambdaContainerParams

Description

Documentation

data LambdaContainerParams

Contains information about a container in which Lambda functions run on Greengrass core devices.

See: newLambdaContainerParams smart constructor.

Constructors

LambdaContainerParams'

Fields

mountROSysfs :: Maybe Bool

Whether or not the container can read information from the device's /sys folder.

Default: false

memorySizeInKB :: Maybe Int

The memory size of the container, expressed in kilobytes.

Default: 16384 (16 MB)

devices :: Maybe [LambdaDeviceMount]

The list of system devices that the container can access.

volumes :: Maybe [LambdaVolumeMount]

The list of volumes that the container can access.

newLambdaContainerParams :: LambdaContainerParams

Create a value of LambdaContainerParams with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:mountROSysfs:LambdaContainerParams', lambdaContainerParams_mountROSysfs - Whether or not the container can read information from the device's /sys folder.

Default: false

$sel:memorySizeInKB:LambdaContainerParams', lambdaContainerParams_memorySizeInKB - The memory size of the container, expressed in kilobytes.

Default: 16384 (16 MB)

$sel:devices:LambdaContainerParams', lambdaContainerParams_devices - The list of system devices that the container can access.

$sel:volumes:LambdaContainerParams', lambdaContainerParams_volumes - The list of volumes that the container can access.

lambdaContainerParams_mountROSysfs :: Lens' LambdaContainerParams (Maybe Bool)

Whether or not the container can read information from the device's /sys folder.

Default: false

lambdaContainerParams_memorySizeInKB :: Lens' LambdaContainerParams (Maybe Int)

The memory size of the container, expressed in kilobytes.

Default: 16384 (16 MB)

lambdaContainerParams_devices :: Lens' LambdaContainerParams (Maybe [LambdaDeviceMount])

The list of system devices that the container can access.

lambdaContainerParams_volumes :: Lens' LambdaContainerParams (Maybe [LambdaVolumeMount])

The list of volumes that the container can access.