Amazonka.GreengrassV2.Types.LambdaEventSource

Description

Documentation

data LambdaEventSource

Contains information about an event source for an Lambda function. The event source defines the topics on which this Lambda function subscribes to receive messages that run the function.

See: newLambdaEventSource smart constructor.

Constructors

LambdaEventSource'

Fields

topic :: Text

The topic to which to subscribe to receive event messages.

type' :: LambdaEventSourceType

The type of event source. Choose from the following options:

PUB_SUB – Subscribe to local publish/subscribe messages. This event source type doesn't support MQTT wildcards (+ and #) in the event source topic.

IOT_CORE – Subscribe to Amazon Web Services IoT Core MQTT messages. This event source type supports MQTT wildcards (+ and #) in the event source topic.

newLambdaEventSource

Arguments

:: Text
$sel:topic:LambdaEventSource'

-> LambdaEventSourceType
$sel:type':LambdaEventSource'

-> LambdaEventSource

Create a value of LambdaEventSource with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:topic:LambdaEventSource', lambdaEventSource_topic - The topic to which to subscribe to receive event messages.

$sel:type':LambdaEventSource', lambdaEventSource_type - The type of event source. Choose from the following options:

PUB_SUB – Subscribe to local publish/subscribe messages. This event source type doesn't support MQTT wildcards (+ and #) in the event source topic.

IOT_CORE – Subscribe to Amazon Web Services IoT Core MQTT messages. This event source type supports MQTT wildcards (+ and #) in the event source topic.

lambdaEventSource_topic :: Lens' LambdaEventSource Text

The topic to which to subscribe to receive event messages.

lambdaEventSource_type :: Lens' LambdaEventSource LambdaEventSourceType

The type of event source. Choose from the following options:

PUB_SUB – Subscribe to local publish/subscribe messages. This event source type doesn't support MQTT wildcards (+ and #) in the event source topic.

IOT_CORE – Subscribe to Amazon Web Services IoT Core MQTT messages. This event source type supports MQTT wildcards (+ and #) in the event source topic.