Amazonka.GroundStation.DeleteDataflowEndpointGroup

Description

Deletes a dataflow endpoint group.

Creating a Request

data DeleteDataflowEndpointGroup

See: newDeleteDataflowEndpointGroup smart constructor.

Constructors

DeleteDataflowEndpointGroup'

Fields

dataflowEndpointGroupId :: Text

UUID of a dataflow endpoint group.

newDeleteDataflowEndpointGroup

Arguments

:: Text
$sel:dataflowEndpointGroupId:DeleteDataflowEndpointGroup'

-> DeleteDataflowEndpointGroup

Create a value of DeleteDataflowEndpointGroup with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:dataflowEndpointGroupId:DeleteDataflowEndpointGroup', deleteDataflowEndpointGroup_dataflowEndpointGroupId - UUID of a dataflow endpoint group.

Request Lenses

deleteDataflowEndpointGroup_dataflowEndpointGroupId :: Lens' DeleteDataflowEndpointGroup Text

UUID of a dataflow endpoint group.

Destructuring the Response

data DataflowEndpointGroupIdResponse

See: newDataflowEndpointGroupIdResponse smart constructor.

Constructors

DataflowEndpointGroupIdResponse'

Fields

dataflowEndpointGroupId :: Maybe Text

UUID of a dataflow endpoint group.

newDataflowEndpointGroupIdResponse :: DataflowEndpointGroupIdResponse

Create a value of DataflowEndpointGroupIdResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:dataflowEndpointGroupId:DataflowEndpointGroupIdResponse', dataflowEndpointGroupIdResponse_dataflowEndpointGroupId - UUID of a dataflow endpoint group.

Response Lenses

dataflowEndpointGroupIdResponse_dataflowEndpointGroupId :: Lens' DataflowEndpointGroupIdResponse (Maybe Text)

UUID of a dataflow endpoint group.