Amazonka.GroundStation.GetMinuteUsage

Description

Returns the number of minutes used by account.

Creating a Request

data GetMinuteUsage

See: newGetMinuteUsage smart constructor.

Constructors

GetMinuteUsage'

Fields

month :: Int

The month being requested, with a value of 1-12.

year :: Int

The year being requested, in the format of YYYY.

newGetMinuteUsage

Arguments

:: Int
$sel:month:GetMinuteUsage'

-> Int
$sel:year:GetMinuteUsage'

-> GetMinuteUsage

Create a value of GetMinuteUsage with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:month:GetMinuteUsage', getMinuteUsage_month - The month being requested, with a value of 1-12.

$sel:year:GetMinuteUsage', getMinuteUsage_year - The year being requested, in the format of YYYY.

Request Lenses

getMinuteUsage_month :: Lens' GetMinuteUsage Int

The month being requested, with a value of 1-12.

getMinuteUsage_year :: Lens' GetMinuteUsage Int

The year being requested, in the format of YYYY.

Destructuring the Response

data GetMinuteUsageResponse

See: newGetMinuteUsageResponse smart constructor.

Constructors

GetMinuteUsageResponse'

Fields

isReservedMinutesCustomer :: Maybe Bool

Returns whether or not an account has signed up for the reserved minutes pricing plan, specific to the month being requested.

upcomingMinutesScheduled :: Maybe Int

Upcoming minutes scheduled for an account, specific to the month being requested.

totalScheduledMinutes :: Maybe Int

Total scheduled minutes for an account, specific to the month being requested.

totalReservedMinuteAllocation :: Maybe Int

Total number of reserved minutes allocated, specific to the month being requested.

estimatedMinutesRemaining :: Maybe Int

Estimated number of minutes remaining for an account, specific to the month being requested.

httpStatus :: Int

The response's http status code.

newGetMinuteUsageResponse

Arguments

:: Int
$sel:httpStatus:GetMinuteUsageResponse'

-> GetMinuteUsageResponse

Create a value of GetMinuteUsageResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:isReservedMinutesCustomer:GetMinuteUsageResponse', getMinuteUsageResponse_isReservedMinutesCustomer - Returns whether or not an account has signed up for the reserved minutes pricing plan, specific to the month being requested.

$sel:upcomingMinutesScheduled:GetMinuteUsageResponse', getMinuteUsageResponse_upcomingMinutesScheduled - Upcoming minutes scheduled for an account, specific to the month being requested.

$sel:totalScheduledMinutes:GetMinuteUsageResponse', getMinuteUsageResponse_totalScheduledMinutes - Total scheduled minutes for an account, specific to the month being requested.

$sel:totalReservedMinuteAllocation:GetMinuteUsageResponse', getMinuteUsageResponse_totalReservedMinuteAllocation - Total number of reserved minutes allocated, specific to the month being requested.

$sel:estimatedMinutesRemaining:GetMinuteUsageResponse', getMinuteUsageResponse_estimatedMinutesRemaining - Estimated number of minutes remaining for an account, specific to the month being requested.

$sel:httpStatus:GetMinuteUsageResponse', getMinuteUsageResponse_httpStatus - The response's http status code.

Response Lenses

getMinuteUsageResponse_isReservedMinutesCustomer :: Lens' GetMinuteUsageResponse (Maybe Bool)

Returns whether or not an account has signed up for the reserved minutes pricing plan, specific to the month being requested.

getMinuteUsageResponse_upcomingMinutesScheduled :: Lens' GetMinuteUsageResponse (Maybe Int)

Upcoming minutes scheduled for an account, specific to the month being requested.

getMinuteUsageResponse_totalScheduledMinutes :: Lens' GetMinuteUsageResponse (Maybe Int)

Total scheduled minutes for an account, specific to the month being requested.

getMinuteUsageResponse_totalReservedMinuteAllocation :: Lens' GetMinuteUsageResponse (Maybe Int)

Total number of reserved minutes allocated, specific to the month being requested.

getMinuteUsageResponse_estimatedMinutesRemaining :: Lens' GetMinuteUsageResponse (Maybe Int)

Estimated number of minutes remaining for an account, specific to the month being requested.

getMinuteUsageResponse_httpStatus :: Lens' GetMinuteUsageResponse Int

The response's http status code.