Amazonka.GroundStation.Types

Description

Service Configuration

defaultService :: Service

API version 2019-05-23 of the Amazon Ground Station SDK configuration.

Errors

_InvalidParameterException :: AsError a => Getting (First ServiceError) a ServiceError

One or more parameters are not valid.

_DependencyException :: AsError a => Getting (First ServiceError) a ServiceError

Dependency encountered an error.

_ResourceLimitExceededException :: AsError a => Getting (First ServiceError) a ServiceError

Account limits for this resource have been exceeded.

_ResourceNotFoundException :: AsError a => Getting (First ServiceError) a ServiceError

Resource was not found.

AngleUnits

newtype AngleUnits

Constructors

AngleUnits'

Fields

fromAngleUnits :: Text

Bundled Patterns

pattern AngleUnits_DEGREE_ANGLE :: AngleUnits

pattern AngleUnits_RADIAN :: AngleUnits

BandwidthUnits

newtype BandwidthUnits

Constructors

BandwidthUnits'

Fields

fromBandwidthUnits :: Text

Bundled Patterns

pattern BandwidthUnits_GHz :: BandwidthUnits

pattern BandwidthUnits_KHz :: BandwidthUnits

pattern BandwidthUnits_MHz :: BandwidthUnits

ConfigCapabilityType

newtype ConfigCapabilityType

Constructors

ConfigCapabilityType'

Fields

fromConfigCapabilityType :: Text

Bundled Patterns

pattern ConfigCapabilityType_Antenna_downlink :: ConfigCapabilityType

pattern ConfigCapabilityType_Antenna_downlink_demod_decode :: ConfigCapabilityType

pattern ConfigCapabilityType_Antenna_uplink :: ConfigCapabilityType

pattern ConfigCapabilityType_Dataflow_endpoint :: ConfigCapabilityType

pattern ConfigCapabilityType_S3_recording :: ConfigCapabilityType

pattern ConfigCapabilityType_Tracking :: ConfigCapabilityType

pattern ConfigCapabilityType_Uplink_echo :: ConfigCapabilityType

ContactStatus

newtype ContactStatus

Constructors

ContactStatus'

Fields

fromContactStatus :: Text

Bundled Patterns

pattern ContactStatus_AVAILABLE :: ContactStatus

pattern ContactStatus_AWS_CANCELLED :: ContactStatus

pattern ContactStatus_AWS_FAILED :: ContactStatus

pattern ContactStatus_CANCELLED :: ContactStatus

pattern ContactStatus_CANCELLING :: ContactStatus

pattern ContactStatus_COMPLETED :: ContactStatus

pattern ContactStatus_FAILED :: ContactStatus

pattern ContactStatus_FAILED_TO_SCHEDULE :: ContactStatus

pattern ContactStatus_PASS :: ContactStatus

pattern ContactStatus_POSTPASS :: ContactStatus

pattern ContactStatus_PREPASS :: ContactStatus

pattern ContactStatus_SCHEDULED :: ContactStatus

pattern ContactStatus_SCHEDULING :: ContactStatus

Criticality

newtype Criticality

Constructors

Criticality'

Fields

fromCriticality :: Text

Bundled Patterns

pattern Criticality_PREFERRED :: Criticality

pattern Criticality_REMOVED :: Criticality

pattern Criticality_REQUIRED :: Criticality

EirpUnits

newtype EirpUnits

Constructors

EirpUnits'

Fields

fromEirpUnits :: Text

Bundled Patterns

pattern EirpUnits_DBW :: EirpUnits

EndpointStatus

newtype EndpointStatus

Constructors

EndpointStatus'

Fields

fromEndpointStatus :: Text

Bundled Patterns

pattern EndpointStatus_Created :: EndpointStatus

pattern EndpointStatus_Creating :: EndpointStatus

pattern EndpointStatus_Deleted :: EndpointStatus

pattern EndpointStatus_Deleting :: EndpointStatus

pattern EndpointStatus_Failed :: EndpointStatus

FrequencyUnits

newtype FrequencyUnits

Constructors

FrequencyUnits'

Fields

fromFrequencyUnits :: Text

Bundled Patterns

pattern FrequencyUnits_GHz :: FrequencyUnits

pattern FrequencyUnits_KHz :: FrequencyUnits

pattern FrequencyUnits_MHz :: FrequencyUnits

Polarization

newtype Polarization

Constructors

Polarization'

Fields

fromPolarization :: Text

Bundled Patterns

pattern Polarization_LEFT_HAND :: Polarization

pattern Polarization_NONE :: Polarization

pattern Polarization_RIGHT_HAND :: Polarization

AntennaDemodDecodeDetails

data AntennaDemodDecodeDetails

Details about an antenna demod decode Config used in a contact.

See: newAntennaDemodDecodeDetails smart constructor.

Constructors

AntennaDemodDecodeDetails'

Fields

outputNode :: Maybe Text

Name of an antenna demod decode output node used in a contact.

newAntennaDemodDecodeDetails :: AntennaDemodDecodeDetails

Create a value of AntennaDemodDecodeDetails with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:outputNode:AntennaDemodDecodeDetails', antennaDemodDecodeDetails_outputNode - Name of an antenna demod decode output node used in a contact.

antennaDemodDecodeDetails_outputNode :: Lens' AntennaDemodDecodeDetails (Maybe Text)

Name of an antenna demod decode output node used in a contact.

AntennaDownlinkConfig

data AntennaDownlinkConfig

Information about how AWS Ground Station should configure an antenna for downlink during a contact.

See: newAntennaDownlinkConfig smart constructor.

Constructors

AntennaDownlinkConfig'

Fields

spectrumConfig :: SpectrumConfig

Object that describes a spectral Config.

newAntennaDownlinkConfig

Arguments

:: SpectrumConfig
$sel:spectrumConfig:AntennaDownlinkConfig'

-> AntennaDownlinkConfig

Create a value of AntennaDownlinkConfig with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:spectrumConfig:AntennaDownlinkConfig', antennaDownlinkConfig_spectrumConfig - Object that describes a spectral Config.

antennaDownlinkConfig_spectrumConfig :: Lens' AntennaDownlinkConfig SpectrumConfig

Object that describes a spectral Config.

AntennaDownlinkDemodDecodeConfig

data AntennaDownlinkDemodDecodeConfig

Information about how AWS Ground Station should conﬁgure an antenna for downlink demod decode during a contact.

See: newAntennaDownlinkDemodDecodeConfig smart constructor.

Constructors

AntennaDownlinkDemodDecodeConfig'

Fields

decodeConfig :: DecodeConfig

Information about the decode Config.

demodulationConfig :: DemodulationConfig

Information about the demodulation Config.

spectrumConfig :: SpectrumConfig

Information about the spectral Config.

newAntennaDownlinkDemodDecodeConfig

Arguments

:: DecodeConfig
$sel:decodeConfig:AntennaDownlinkDemodDecodeConfig'

-> DemodulationConfig
$sel:demodulationConfig:AntennaDownlinkDemodDecodeConfig'

-> SpectrumConfig
$sel:spectrumConfig:AntennaDownlinkDemodDecodeConfig'

-> AntennaDownlinkDemodDecodeConfig

Create a value of AntennaDownlinkDemodDecodeConfig with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:decodeConfig:AntennaDownlinkDemodDecodeConfig', antennaDownlinkDemodDecodeConfig_decodeConfig - Information about the decode Config.

$sel:demodulationConfig:AntennaDownlinkDemodDecodeConfig', antennaDownlinkDemodDecodeConfig_demodulationConfig - Information about the demodulation Config.

$sel:spectrumConfig:AntennaDownlinkDemodDecodeConfig', antennaDownlinkDemodDecodeConfig_spectrumConfig - Information about the spectral Config.

antennaDownlinkDemodDecodeConfig_decodeConfig :: Lens' AntennaDownlinkDemodDecodeConfig DecodeConfig

Information about the decode Config.

antennaDownlinkDemodDecodeConfig_demodulationConfig :: Lens' AntennaDownlinkDemodDecodeConfig DemodulationConfig

Information about the demodulation Config.

antennaDownlinkDemodDecodeConfig_spectrumConfig :: Lens' AntennaDownlinkDemodDecodeConfig SpectrumConfig

Information about the spectral Config.

AntennaUplinkConfig

data AntennaUplinkConfig

Information about the uplink Config of an antenna.

See: newAntennaUplinkConfig smart constructor.

Constructors

AntennaUplinkConfig'

Fields

transmitDisabled :: Maybe Bool

Whether or not uplink transmit is disabled.

spectrumConfig :: UplinkSpectrumConfig

Information about the uplink spectral Config.

targetEirp :: Eirp

EIRP of the target.

newAntennaUplinkConfig

Arguments

:: UplinkSpectrumConfig
$sel:spectrumConfig:AntennaUplinkConfig'

-> Eirp
$sel:targetEirp:AntennaUplinkConfig'

-> AntennaUplinkConfig

Create a value of AntennaUplinkConfig with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:transmitDisabled:AntennaUplinkConfig', antennaUplinkConfig_transmitDisabled - Whether or not uplink transmit is disabled.

$sel:spectrumConfig:AntennaUplinkConfig', antennaUplinkConfig_spectrumConfig - Information about the uplink spectral Config.

$sel:targetEirp:AntennaUplinkConfig', antennaUplinkConfig_targetEirp - EIRP of the target.

antennaUplinkConfig_transmitDisabled :: Lens' AntennaUplinkConfig (Maybe Bool)

Whether or not uplink transmit is disabled.

antennaUplinkConfig_spectrumConfig :: Lens' AntennaUplinkConfig UplinkSpectrumConfig

Information about the uplink spectral Config.

antennaUplinkConfig_targetEirp :: Lens' AntennaUplinkConfig Eirp

EIRP of the target.

ConfigDetails

data ConfigDetails

Details for certain Config object types in a contact.

See: newConfigDetails smart constructor.

Constructors

ConfigDetails'

Fields

antennaDemodDecodeDetails :: Maybe AntennaDemodDecodeDetails

Details for antenna demod decode Config in a contact.

s3RecordingDetails :: Maybe S3RecordingDetails

Details for an S3 recording Config in a contact.

endpointDetails :: Maybe EndpointDetails

newConfigDetails :: ConfigDetails

Create a value of ConfigDetails with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:antennaDemodDecodeDetails:ConfigDetails', configDetails_antennaDemodDecodeDetails - Details for antenna demod decode Config in a contact.

$sel:s3RecordingDetails:ConfigDetails', configDetails_s3RecordingDetails - Details for an S3 recording Config in a contact.

$sel:endpointDetails:ConfigDetails', configDetails_endpointDetails - Undocumented member.

configDetails_antennaDemodDecodeDetails :: Lens' ConfigDetails (Maybe AntennaDemodDecodeDetails)

Details for antenna demod decode Config in a contact.

configDetails_s3RecordingDetails :: Lens' ConfigDetails (Maybe S3RecordingDetails)

Details for an S3 recording Config in a contact.

configDetails_endpointDetails :: Lens' ConfigDetails (Maybe EndpointDetails)

Undocumented member.

ConfigIdResponse

data ConfigIdResponse

See: newConfigIdResponse smart constructor.

Constructors

ConfigIdResponse'

Fields

configArn :: Maybe Text

ARN of a Config.

configId :: Maybe Text

UUID of a Config.

configType :: Maybe ConfigCapabilityType

Type of a Config.

newConfigIdResponse :: ConfigIdResponse

Create a value of ConfigIdResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:configArn:ConfigIdResponse', configIdResponse_configArn - ARN of a Config.

$sel:configId:ConfigIdResponse', configIdResponse_configId - UUID of a Config.

$sel:configType:ConfigIdResponse', configIdResponse_configType - Type of a Config.

configIdResponse_configArn :: Lens' ConfigIdResponse (Maybe Text)

ARN of a Config.

configIdResponse_configId :: Lens' ConfigIdResponse (Maybe Text)

UUID of a Config.

configIdResponse_configType :: Lens' ConfigIdResponse (Maybe ConfigCapabilityType)

Type of a Config.

ConfigListItem

data ConfigListItem

An item in a list of Config objects.

See: newConfigListItem smart constructor.

Constructors

ConfigListItem'

Fields

configArn :: Maybe Text

ARN of a Config.

name :: Maybe Text

Name of a Config.

configId :: Maybe Text

UUID of a Config.

configType :: Maybe ConfigCapabilityType

Type of a Config.

newConfigListItem :: ConfigListItem

Create a value of ConfigListItem with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:configArn:ConfigListItem', configListItem_configArn - ARN of a Config.

$sel:name:ConfigListItem', configListItem_name - Name of a Config.

$sel:configId:ConfigListItem', configListItem_configId - UUID of a Config.

$sel:configType:ConfigListItem', configListItem_configType - Type of a Config.

configListItem_configArn :: Lens' ConfigListItem (Maybe Text)

ARN of a Config.

configListItem_name :: Lens' ConfigListItem (Maybe Text)

Name of a Config.

configListItem_configId :: Lens' ConfigListItem (Maybe Text)

UUID of a Config.

configListItem_configType :: Lens' ConfigListItem (Maybe ConfigCapabilityType)

Type of a Config.

ConfigTypeData

data ConfigTypeData

Object containing the parameters of a Config.

See the subtype definitions for what each type of Config contains.

See: newConfigTypeData smart constructor.

Constructors

ConfigTypeData'

Fields

antennaDownlinkDemodDecodeConfig :: Maybe AntennaDownlinkDemodDecodeConfig

Information about how AWS Ground Station should conﬁgure an antenna for downlink demod decode during a contact.

antennaDownlinkConfig :: Maybe AntennaDownlinkConfig

Information about how AWS Ground Station should configure an antenna for downlink during a contact.

antennaUplinkConfig :: Maybe AntennaUplinkConfig

Information about how AWS Ground Station should conﬁgure an antenna for uplink during a contact.

uplinkEchoConfig :: Maybe UplinkEchoConfig

Information about an uplink echo Config.

Parameters from the AntennaUplinkConfig, corresponding to the specified AntennaUplinkConfigArn, are used when this UplinkEchoConfig is used in a contact.

s3RecordingConfig :: Maybe S3RecordingConfig

Information about an S3 recording Config.

trackingConfig :: Maybe TrackingConfig

Object that determines whether tracking should be used during a contact executed with this Config in the mission profile.

dataflowEndpointConfig :: Maybe DataflowEndpointConfig

Information about the dataflow endpoint Config.

newConfigTypeData :: ConfigTypeData

Create a value of ConfigTypeData with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:antennaDownlinkDemodDecodeConfig:ConfigTypeData', configTypeData_antennaDownlinkDemodDecodeConfig - Information about how AWS Ground Station should conﬁgure an antenna for downlink demod decode during a contact.

$sel:antennaDownlinkConfig:ConfigTypeData', configTypeData_antennaDownlinkConfig - Information about how AWS Ground Station should configure an antenna for downlink during a contact.

$sel:antennaUplinkConfig:ConfigTypeData', configTypeData_antennaUplinkConfig - Information about how AWS Ground Station should conﬁgure an antenna for uplink during a contact.

$sel:uplinkEchoConfig:ConfigTypeData', configTypeData_uplinkEchoConfig - Information about an uplink echo Config.

Parameters from the AntennaUplinkConfig, corresponding to the specified AntennaUplinkConfigArn, are used when this UplinkEchoConfig is used in a contact.

$sel:s3RecordingConfig:ConfigTypeData', configTypeData_s3RecordingConfig - Information about an S3 recording Config.

$sel:trackingConfig:ConfigTypeData', configTypeData_trackingConfig - Object that determines whether tracking should be used during a contact executed with this Config in the mission profile.

$sel:dataflowEndpointConfig:ConfigTypeData', configTypeData_dataflowEndpointConfig - Information about the dataflow endpoint Config.

configTypeData_antennaDownlinkDemodDecodeConfig :: Lens' ConfigTypeData (Maybe AntennaDownlinkDemodDecodeConfig)

Information about how AWS Ground Station should conﬁgure an antenna for downlink demod decode during a contact.

configTypeData_antennaDownlinkConfig :: Lens' ConfigTypeData (Maybe AntennaDownlinkConfig)

Information about how AWS Ground Station should configure an antenna for downlink during a contact.

configTypeData_antennaUplinkConfig :: Lens' ConfigTypeData (Maybe AntennaUplinkConfig)

Information about how AWS Ground Station should conﬁgure an antenna for uplink during a contact.

configTypeData_uplinkEchoConfig :: Lens' ConfigTypeData (Maybe UplinkEchoConfig)

Information about an uplink echo Config.

Parameters from the AntennaUplinkConfig, corresponding to the specified AntennaUplinkConfigArn, are used when this UplinkEchoConfig is used in a contact.

configTypeData_s3RecordingConfig :: Lens' ConfigTypeData (Maybe S3RecordingConfig)

Information about an S3 recording Config.

configTypeData_trackingConfig :: Lens' ConfigTypeData (Maybe TrackingConfig)

Object that determines whether tracking should be used during a contact executed with this Config in the mission profile.

configTypeData_dataflowEndpointConfig :: Lens' ConfigTypeData (Maybe DataflowEndpointConfig)

Information about the dataflow endpoint Config.

ContactData

data ContactData

Data describing a contact.

See: newContactData smart constructor.

Constructors

ContactData'

Fields

contactStatus :: Maybe ContactStatus

Status of a contact.

missionProfileArn :: Maybe Text

ARN of a mission profile.

startTime :: Maybe POSIX

Start time of a contact.

satelliteArn :: Maybe Text

ARN of a satellite.

maximumElevation :: Maybe Elevation

Maximum elevation angle of a contact.

groundStation :: Maybe Text

Name of a ground station.

endTime :: Maybe POSIX

End time of a contact.

contactId :: Maybe Text

UUID of a contact.

region :: Maybe Text

Region of a contact.

postPassEndTime :: Maybe POSIX

Amount of time after a contact ends that you’d like to receive a CloudWatch event indicating the pass has finished.

prePassStartTime :: Maybe POSIX

Amount of time prior to contact start you’d like to receive a CloudWatch event indicating an upcoming pass.

errorMessage :: Maybe Text

Error message of a contact.

tags :: Maybe (HashMap Text Text)

Tags assigned to a contact.

newContactData :: ContactData

Create a value of ContactData with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:contactStatus:ContactData', contactData_contactStatus - Status of a contact.

$sel:missionProfileArn:ContactData', contactData_missionProfileArn - ARN of a mission profile.

$sel:startTime:ContactData', contactData_startTime - Start time of a contact.

$sel:satelliteArn:ContactData', contactData_satelliteArn - ARN of a satellite.

$sel:maximumElevation:ContactData', contactData_maximumElevation - Maximum elevation angle of a contact.

$sel:groundStation:ContactData', contactData_groundStation - Name of a ground station.

$sel:endTime:ContactData', contactData_endTime - End time of a contact.

$sel:contactId:ContactData', contactData_contactId - UUID of a contact.

$sel:region:ContactData', contactData_region - Region of a contact.

$sel:postPassEndTime:ContactData', contactData_postPassEndTime - Amount of time after a contact ends that you’d like to receive a CloudWatch event indicating the pass has finished.

$sel:prePassStartTime:ContactData', contactData_prePassStartTime - Amount of time prior to contact start you’d like to receive a CloudWatch event indicating an upcoming pass.

$sel:errorMessage:ContactData', contactData_errorMessage - Error message of a contact.

$sel:tags:ContactData', contactData_tags - Tags assigned to a contact.

contactData_contactStatus :: Lens' ContactData (Maybe ContactStatus)

Status of a contact.

contactData_missionProfileArn :: Lens' ContactData (Maybe Text)

ARN of a mission profile.

contactData_startTime :: Lens' ContactData (Maybe UTCTime)

Start time of a contact.

contactData_satelliteArn :: Lens' ContactData (Maybe Text)

ARN of a satellite.

contactData_maximumElevation :: Lens' ContactData (Maybe Elevation)

Maximum elevation angle of a contact.

contactData_groundStation :: Lens' ContactData (Maybe Text)

Name of a ground station.

contactData_endTime :: Lens' ContactData (Maybe UTCTime)

End time of a contact.

contactData_contactId :: Lens' ContactData (Maybe Text)

UUID of a contact.

contactData_region :: Lens' ContactData (Maybe Text)

Region of a contact.

contactData_postPassEndTime :: Lens' ContactData (Maybe UTCTime)

Amount of time after a contact ends that you’d like to receive a CloudWatch event indicating the pass has finished.

contactData_prePassStartTime :: Lens' ContactData (Maybe UTCTime)

Amount of time prior to contact start you’d like to receive a CloudWatch event indicating an upcoming pass.

contactData_errorMessage :: Lens' ContactData (Maybe Text)

Error message of a contact.

contactData_tags :: Lens' ContactData (Maybe (HashMap Text Text))

Tags assigned to a contact.

ContactIdResponse

data ContactIdResponse

See: newContactIdResponse smart constructor.

Constructors

ContactIdResponse'

Fields

contactId :: Maybe Text

UUID of a contact.

newContactIdResponse :: ContactIdResponse

Create a value of ContactIdResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:contactId:ContactIdResponse', contactIdResponse_contactId - UUID of a contact.

contactIdResponse_contactId :: Lens' ContactIdResponse (Maybe Text)

UUID of a contact.

DataflowDetail

data DataflowDetail

Information about a dataflow edge used in a contact.

See: newDataflowDetail smart constructor.

Constructors

DataflowDetail'

Fields

destination :: Maybe Destination

source :: Maybe Source

errorMessage :: Maybe Text

Error message for a dataflow.

newDataflowDetail :: DataflowDetail

Create a value of DataflowDetail with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:destination:DataflowDetail', dataflowDetail_destination - Undocumented member.

$sel:source:DataflowDetail', dataflowDetail_source - Undocumented member.

$sel:errorMessage:DataflowDetail', dataflowDetail_errorMessage - Error message for a dataflow.

dataflowDetail_destination :: Lens' DataflowDetail (Maybe Destination)

Undocumented member.

dataflowDetail_source :: Lens' DataflowDetail (Maybe Source)

Undocumented member.

dataflowDetail_errorMessage :: Lens' DataflowDetail (Maybe Text)

Error message for a dataflow.

DataflowEndpoint

data DataflowEndpoint

Information about a dataflow endpoint.

See: newDataflowEndpoint smart constructor.

Constructors

DataflowEndpoint'

Fields

mtu :: Maybe Natural

Maximum transmission unit (MTU) size in bytes of a dataflow endpoint.

status :: Maybe EndpointStatus

Status of a dataflow endpoint.

address :: Maybe SocketAddress

Socket address of a dataflow endpoint.

name :: Maybe Text

Name of a dataflow endpoint.

newDataflowEndpoint :: DataflowEndpoint

Create a value of DataflowEndpoint with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:mtu:DataflowEndpoint', dataflowEndpoint_mtu - Maximum transmission unit (MTU) size in bytes of a dataflow endpoint.

$sel:status:DataflowEndpoint', dataflowEndpoint_status - Status of a dataflow endpoint.

$sel:address:DataflowEndpoint', dataflowEndpoint_address - Socket address of a dataflow endpoint.

$sel:name:DataflowEndpoint', dataflowEndpoint_name - Name of a dataflow endpoint.

dataflowEndpoint_mtu :: Lens' DataflowEndpoint (Maybe Natural)

Maximum transmission unit (MTU) size in bytes of a dataflow endpoint.

dataflowEndpoint_status :: Lens' DataflowEndpoint (Maybe EndpointStatus)

Status of a dataflow endpoint.

dataflowEndpoint_address :: Lens' DataflowEndpoint (Maybe SocketAddress)

Socket address of a dataflow endpoint.

dataflowEndpoint_name :: Lens' DataflowEndpoint (Maybe Text)

Name of a dataflow endpoint.

DataflowEndpointConfig

data DataflowEndpointConfig

Information about the dataflow endpoint Config.

See: newDataflowEndpointConfig smart constructor.

Constructors

DataflowEndpointConfig'

Fields

dataflowEndpointRegion :: Maybe Text

Region of a dataflow endpoint.

dataflowEndpointName :: Text

Name of a dataflow endpoint.

newDataflowEndpointConfig

Arguments

:: Text
$sel:dataflowEndpointName:DataflowEndpointConfig'

-> DataflowEndpointConfig

Create a value of DataflowEndpointConfig with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:dataflowEndpointRegion:DataflowEndpointConfig', dataflowEndpointConfig_dataflowEndpointRegion - Region of a dataflow endpoint.

$sel:dataflowEndpointName:DataflowEndpointConfig', dataflowEndpointConfig_dataflowEndpointName - Name of a dataflow endpoint.

dataflowEndpointConfig_dataflowEndpointRegion :: Lens' DataflowEndpointConfig (Maybe Text)

Region of a dataflow endpoint.

dataflowEndpointConfig_dataflowEndpointName :: Lens' DataflowEndpointConfig Text

Name of a dataflow endpoint.

DataflowEndpointGroupIdResponse

data DataflowEndpointGroupIdResponse

See: newDataflowEndpointGroupIdResponse smart constructor.

Constructors

DataflowEndpointGroupIdResponse'

Fields

dataflowEndpointGroupId :: Maybe Text

UUID of a dataflow endpoint group.

newDataflowEndpointGroupIdResponse :: DataflowEndpointGroupIdResponse

Create a value of DataflowEndpointGroupIdResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:dataflowEndpointGroupId:DataflowEndpointGroupIdResponse', dataflowEndpointGroupIdResponse_dataflowEndpointGroupId - UUID of a dataflow endpoint group.

dataflowEndpointGroupIdResponse_dataflowEndpointGroupId :: Lens' DataflowEndpointGroupIdResponse (Maybe Text)

UUID of a dataflow endpoint group.

DataflowEndpointListItem

data DataflowEndpointListItem

Item in a list of DataflowEndpoint groups.

See: newDataflowEndpointListItem smart constructor.

Constructors

DataflowEndpointListItem'

Fields

dataflowEndpointGroupArn :: Maybe Text

ARN of a dataflow endpoint group.

dataflowEndpointGroupId :: Maybe Text

UUID of a dataflow endpoint group.

newDataflowEndpointListItem :: DataflowEndpointListItem

Create a value of DataflowEndpointListItem with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:dataflowEndpointGroupArn:DataflowEndpointListItem', dataflowEndpointListItem_dataflowEndpointGroupArn - ARN of a dataflow endpoint group.

$sel:dataflowEndpointGroupId:DataflowEndpointListItem', dataflowEndpointListItem_dataflowEndpointGroupId - UUID of a dataflow endpoint group.

dataflowEndpointListItem_dataflowEndpointGroupArn :: Lens' DataflowEndpointListItem (Maybe Text)

ARN of a dataflow endpoint group.

dataflowEndpointListItem_dataflowEndpointGroupId :: Lens' DataflowEndpointListItem (Maybe Text)

UUID of a dataflow endpoint group.

DecodeConfig

data DecodeConfig

Information about the decode Config.

See: newDecodeConfig smart constructor.

Constructors

DecodeConfig'

Fields

unvalidatedJSON :: Text

Unvalidated JSON of a decode Config.

newDecodeConfig

Arguments

:: Text
$sel:unvalidatedJSON:DecodeConfig'

-> DecodeConfig

Create a value of DecodeConfig with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:unvalidatedJSON:DecodeConfig', decodeConfig_unvalidatedJSON - Unvalidated JSON of a decode Config.

decodeConfig_unvalidatedJSON :: Lens' DecodeConfig Text

Unvalidated JSON of a decode Config.

DemodulationConfig

data DemodulationConfig

Information about the demodulation Config.

See: newDemodulationConfig smart constructor.

Constructors

DemodulationConfig'

Fields

unvalidatedJSON :: Text

Unvalidated JSON of a demodulation Config.

newDemodulationConfig

Arguments

:: Text
$sel:unvalidatedJSON:DemodulationConfig'

-> DemodulationConfig

Create a value of DemodulationConfig with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:unvalidatedJSON:DemodulationConfig', demodulationConfig_unvalidatedJSON - Unvalidated JSON of a demodulation Config.

demodulationConfig_unvalidatedJSON :: Lens' DemodulationConfig Text

Unvalidated JSON of a demodulation Config.

Destination

data Destination

Dataflow details for the destination side.

See: newDestination smart constructor.

Constructors

Destination'

Fields

dataflowDestinationRegion :: Maybe Text

Region of a dataflow destination.

configId :: Maybe Text

UUID of a Config.

configType :: Maybe ConfigCapabilityType

Type of a Config.

configDetails :: Maybe ConfigDetails

Additional details for a Config, if type is dataflow endpoint or antenna demod decode.

newDestination :: Destination

Create a value of Destination with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:dataflowDestinationRegion:Destination', destination_dataflowDestinationRegion - Region of a dataflow destination.

$sel:configId:Destination', destination_configId - UUID of a Config.

$sel:configType:Destination', destination_configType - Type of a Config.

$sel:configDetails:Destination', destination_configDetails - Additional details for a Config, if type is dataflow endpoint or antenna demod decode.

destination_dataflowDestinationRegion :: Lens' Destination (Maybe Text)

Region of a dataflow destination.

destination_configId :: Lens' Destination (Maybe Text)

UUID of a Config.

destination_configType :: Lens' Destination (Maybe ConfigCapabilityType)

Type of a Config.

destination_configDetails :: Lens' Destination (Maybe ConfigDetails)

Additional details for a Config, if type is dataflow endpoint or antenna demod decode.

Eirp

data Eirp

Object that represents EIRP.

See: newEirp smart constructor.

Constructors

Eirp'

Fields

units :: EirpUnits

Units of an EIRP.

value :: Double

Value of an EIRP. Valid values are between 20.0 to 50.0 dBW.

newEirp

Arguments

:: EirpUnits
$sel:units:Eirp'

-> Double
$sel:value:Eirp'

-> Eirp

Create a value of Eirp with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:units:Eirp', eirp_units - Units of an EIRP.

$sel:value:Eirp', eirp_value - Value of an EIRP. Valid values are between 20.0 to 50.0 dBW.

eirp_units :: Lens' Eirp EirpUnits

Units of an EIRP.

eirp_value :: Lens' Eirp Double

Value of an EIRP. Valid values are between 20.0 to 50.0 dBW.

Elevation

data Elevation

Elevation angle of the satellite in the sky during a contact.

See: newElevation smart constructor.

Constructors

Elevation'

Fields

unit :: AngleUnits

Elevation angle units.

value :: Double

Elevation angle value.

newElevation

Arguments

:: AngleUnits
$sel:unit:Elevation'

-> Double
$sel:value:Elevation'

-> Elevation

Create a value of Elevation with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:unit:Elevation', elevation_unit - Elevation angle units.

$sel:value:Elevation', elevation_value - Elevation angle value.

elevation_unit :: Lens' Elevation AngleUnits

Elevation angle units.

elevation_value :: Lens' Elevation Double

Elevation angle value.

EndpointDetails

data EndpointDetails

Information about the endpoint details.

See: newEndpointDetails smart constructor.

Constructors

EndpointDetails'

Fields

securityDetails :: Maybe SecurityDetails

Endpoint security details.

endpoint :: Maybe DataflowEndpoint

A dataflow endpoint.

newEndpointDetails :: EndpointDetails

Create a value of EndpointDetails with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:securityDetails:EndpointDetails', endpointDetails_securityDetails - Endpoint security details.

$sel:endpoint:EndpointDetails', endpointDetails_endpoint - A dataflow endpoint.

endpointDetails_securityDetails :: Lens' EndpointDetails (Maybe SecurityDetails)

Endpoint security details.

endpointDetails_endpoint :: Lens' EndpointDetails (Maybe DataflowEndpoint)

A dataflow endpoint.

Frequency

data Frequency

Object that describes the frequency.

See: newFrequency smart constructor.

Constructors

Frequency'

Fields

units :: FrequencyUnits

Frequency units.

value :: Double

Frequency value. Valid values are between 2200 to 2300 MHz and 7750 to 8400 MHz for downlink and 2025 to 2120 MHz for uplink.

newFrequency

Arguments

:: FrequencyUnits
$sel:units:Frequency'

-> Double
$sel:value:Frequency'

-> Frequency

Create a value of Frequency with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:units:Frequency', frequency_units - Frequency units.

$sel:value:Frequency', frequency_value - Frequency value. Valid values are between 2200 to 2300 MHz and 7750 to 8400 MHz for downlink and 2025 to 2120 MHz for uplink.

frequency_units :: Lens' Frequency FrequencyUnits

Frequency units.

frequency_value :: Lens' Frequency Double

Frequency value. Valid values are between 2200 to 2300 MHz and 7750 to 8400 MHz for downlink and 2025 to 2120 MHz for uplink.

FrequencyBandwidth

data FrequencyBandwidth

Object that describes the frequency bandwidth.

See: newFrequencyBandwidth smart constructor.

Constructors

FrequencyBandwidth'

Fields

units :: BandwidthUnits

Frequency bandwidth units.

value :: Double

Frequency bandwidth value. AWS Ground Station currently has the following bandwidth limitations:

For AntennaDownlinkDemodDecodeconfig, valid values are between 125 kHz to 650 MHz.

For AntennaDownlinkconfig, valid values are between 10 kHz to 54 MHz.

For AntennaUplinkConfig, valid values are between 10 kHz to 54 MHz.

newFrequencyBandwidth

Arguments

:: BandwidthUnits
$sel:units:FrequencyBandwidth'

-> Double
$sel:value:FrequencyBandwidth'

-> FrequencyBandwidth

Create a value of FrequencyBandwidth with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:units:FrequencyBandwidth', frequencyBandwidth_units - Frequency bandwidth units.

$sel:value:FrequencyBandwidth', frequencyBandwidth_value - Frequency bandwidth value. AWS Ground Station currently has the following bandwidth limitations:

For AntennaDownlinkDemodDecodeconfig, valid values are between 125 kHz to 650 MHz.

For AntennaDownlinkconfig, valid values are between 10 kHz to 54 MHz.

For AntennaUplinkConfig, valid values are between 10 kHz to 54 MHz.

frequencyBandwidth_units :: Lens' FrequencyBandwidth BandwidthUnits

Frequency bandwidth units.

frequencyBandwidth_value :: Lens' FrequencyBandwidth Double

Frequency bandwidth value. AWS Ground Station currently has the following bandwidth limitations:

For AntennaDownlinkDemodDecodeconfig, valid values are between 125 kHz to 650 MHz.

For AntennaDownlinkconfig, valid values are between 10 kHz to 54 MHz.

For AntennaUplinkConfig, valid values are between 10 kHz to 54 MHz.

GroundStationData

data GroundStationData

Information about the ground station data.

See: newGroundStationData smart constructor.

Constructors

GroundStationData'

Fields

groundStationId :: Maybe Text

UUID of a ground station.

groundStationName :: Maybe Text

Name of a ground station.

region :: Maybe Text

Ground station Region.

newGroundStationData :: GroundStationData

Create a value of GroundStationData with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:groundStationId:GroundStationData', groundStationData_groundStationId - UUID of a ground station.

$sel:groundStationName:GroundStationData', groundStationData_groundStationName - Name of a ground station.

$sel:region:GroundStationData', groundStationData_region - Ground station Region.

groundStationData_groundStationId :: Lens' GroundStationData (Maybe Text)

UUID of a ground station.

groundStationData_groundStationName :: Lens' GroundStationData (Maybe Text)

Name of a ground station.

groundStationData_region :: Lens' GroundStationData (Maybe Text)

Ground station Region.

MissionProfileIdResponse

data MissionProfileIdResponse

See: newMissionProfileIdResponse smart constructor.

Constructors

MissionProfileIdResponse'

Fields

missionProfileId :: Maybe Text

UUID of a mission profile.

newMissionProfileIdResponse :: MissionProfileIdResponse

Create a value of MissionProfileIdResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:missionProfileId:MissionProfileIdResponse', missionProfileIdResponse_missionProfileId - UUID of a mission profile.

missionProfileIdResponse_missionProfileId :: Lens' MissionProfileIdResponse (Maybe Text)

UUID of a mission profile.

MissionProfileListItem

data MissionProfileListItem

Item in a list of mission profiles.

See: newMissionProfileListItem smart constructor.

Constructors

MissionProfileListItem'

Fields

missionProfileId :: Maybe Text

UUID of a mission profile.

missionProfileArn :: Maybe Text

ARN of a mission profile.

name :: Maybe Text

Name of a mission profile.

region :: Maybe Text

Region of a mission profile.

newMissionProfileListItem :: MissionProfileListItem

Create a value of MissionProfileListItem with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:missionProfileId:MissionProfileListItem', missionProfileListItem_missionProfileId - UUID of a mission profile.

$sel:missionProfileArn:MissionProfileListItem', missionProfileListItem_missionProfileArn - ARN of a mission profile.

$sel:name:MissionProfileListItem', missionProfileListItem_name - Name of a mission profile.

$sel:region:MissionProfileListItem', missionProfileListItem_region - Region of a mission profile.

missionProfileListItem_missionProfileId :: Lens' MissionProfileListItem (Maybe Text)

UUID of a mission profile.

missionProfileListItem_missionProfileArn :: Lens' MissionProfileListItem (Maybe Text)

ARN of a mission profile.

missionProfileListItem_name :: Lens' MissionProfileListItem (Maybe Text)

Name of a mission profile.

missionProfileListItem_region :: Lens' MissionProfileListItem (Maybe Text)

Region of a mission profile.

S3RecordingConfig

data S3RecordingConfig

Information about an S3 recording Config.

See: newS3RecordingConfig smart constructor.

Constructors

S3RecordingConfig'

Fields

prefix :: Maybe Text

S3 Key prefix to prefice data files.

bucketArn :: Text

ARN of the bucket to record to.

roleArn :: Text

ARN of the role Ground Station assumes to write data to the bucket.

newS3RecordingConfig

Arguments

:: Text
$sel:bucketArn:S3RecordingConfig'

-> Text
$sel:roleArn:S3RecordingConfig'

-> S3RecordingConfig

Create a value of S3RecordingConfig with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:prefix:S3RecordingConfig', s3RecordingConfig_prefix - S3 Key prefix to prefice data files.

$sel:bucketArn:S3RecordingConfig', s3RecordingConfig_bucketArn - ARN of the bucket to record to.

$sel:roleArn:S3RecordingConfig', s3RecordingConfig_roleArn - ARN of the role Ground Station assumes to write data to the bucket.

s3RecordingConfig_prefix :: Lens' S3RecordingConfig (Maybe Text)

S3 Key prefix to prefice data files.

s3RecordingConfig_bucketArn :: Lens' S3RecordingConfig Text

ARN of the bucket to record to.

s3RecordingConfig_roleArn :: Lens' S3RecordingConfig Text

ARN of the role Ground Station assumes to write data to the bucket.

S3RecordingDetails

data S3RecordingDetails

Details about an S3 recording Config used in a contact.

See: newS3RecordingDetails smart constructor.

Constructors

S3RecordingDetails'

Fields

keyTemplate :: Maybe Text

Template of the S3 key used.

bucketArn :: Maybe Text

ARN of the bucket used.

newS3RecordingDetails :: S3RecordingDetails

Create a value of S3RecordingDetails with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:keyTemplate:S3RecordingDetails', s3RecordingDetails_keyTemplate - Template of the S3 key used.

$sel:bucketArn:S3RecordingDetails', s3RecordingDetails_bucketArn - ARN of the bucket used.

s3RecordingDetails_keyTemplate :: Lens' S3RecordingDetails (Maybe Text)

Template of the S3 key used.

s3RecordingDetails_bucketArn :: Lens' S3RecordingDetails (Maybe Text)

ARN of the bucket used.

SatelliteListItem

data SatelliteListItem

Item in a list of satellites.

See: newSatelliteListItem smart constructor.

Constructors

SatelliteListItem'

Fields

satelliteId :: Maybe Text

UUID of a satellite.

satelliteArn :: Maybe Text

ARN of a satellite.

groundStations :: Maybe [Text]

A list of ground stations to which the satellite is on-boarded.

noradSatelliteID :: Maybe Natural

NORAD satellite ID number.

newSatelliteListItem :: SatelliteListItem

Create a value of SatelliteListItem with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:satelliteId:SatelliteListItem', satelliteListItem_satelliteId - UUID of a satellite.

$sel:satelliteArn:SatelliteListItem', satelliteListItem_satelliteArn - ARN of a satellite.

$sel:groundStations:SatelliteListItem', satelliteListItem_groundStations - A list of ground stations to which the satellite is on-boarded.

$sel:noradSatelliteID:SatelliteListItem', satelliteListItem_noradSatelliteID - NORAD satellite ID number.

satelliteListItem_satelliteId :: Lens' SatelliteListItem (Maybe Text)

UUID of a satellite.

satelliteListItem_satelliteArn :: Lens' SatelliteListItem (Maybe Text)

ARN of a satellite.

satelliteListItem_groundStations :: Lens' SatelliteListItem (Maybe [Text])

A list of ground stations to which the satellite is on-boarded.

satelliteListItem_noradSatelliteID :: Lens' SatelliteListItem (Maybe Natural)

NORAD satellite ID number.

SecurityDetails

data SecurityDetails

Information about endpoints.

See: newSecurityDetails smart constructor.

Constructors

SecurityDetails'

Fields

roleArn :: Text

ARN to a role needed for connecting streams to your instances.

securityGroupIds :: [Text]

The security groups to attach to the elastic network interfaces.

subnetIds :: [Text]

A list of subnets where AWS Ground Station places elastic network interfaces to send streams to your instances.

newSecurityDetails

Arguments

:: Text
$sel:roleArn:SecurityDetails'

-> SecurityDetails

Create a value of SecurityDetails with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:roleArn:SecurityDetails', securityDetails_roleArn - ARN to a role needed for connecting streams to your instances.

$sel:securityGroupIds:SecurityDetails', securityDetails_securityGroupIds - The security groups to attach to the elastic network interfaces.

$sel:subnetIds:SecurityDetails', securityDetails_subnetIds - A list of subnets where AWS Ground Station places elastic network interfaces to send streams to your instances.

securityDetails_roleArn :: Lens' SecurityDetails Text

ARN to a role needed for connecting streams to your instances.

securityDetails_securityGroupIds :: Lens' SecurityDetails [Text]

The security groups to attach to the elastic network interfaces.

securityDetails_subnetIds :: Lens' SecurityDetails [Text]

A list of subnets where AWS Ground Station places elastic network interfaces to send streams to your instances.

SocketAddress

data SocketAddress

Information about the socket address.

See: newSocketAddress smart constructor.

Constructors

SocketAddress'

Fields

name :: Text

Name of a socket address.

port :: Int

Port of a socket address.

newSocketAddress

Arguments

:: Text
$sel:name:SocketAddress'

-> Int
$sel:port:SocketAddress'

-> SocketAddress

Create a value of SocketAddress with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:name:SocketAddress', socketAddress_name - Name of a socket address.

$sel:port:SocketAddress', socketAddress_port - Port of a socket address.

socketAddress_name :: Lens' SocketAddress Text

Name of a socket address.

socketAddress_port :: Lens' SocketAddress Int

Port of a socket address.

Source

data Source

Dataflow details for the source side.

See: newSource smart constructor.

Constructors

Source'

Fields

dataflowSourceRegion :: Maybe Text

Region of a dataflow source.

configId :: Maybe Text

UUID of a Config.

configType :: Maybe ConfigCapabilityType

Type of a Config.

configDetails :: Maybe ConfigDetails

Additional details for a Config, if type is dataflow endpoint or antenna demod decode.

newSource :: Source

Create a value of Source with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:dataflowSourceRegion:Source', source_dataflowSourceRegion - Region of a dataflow source.

$sel:configId:Source', source_configId - UUID of a Config.

$sel:configType:Source', source_configType - Type of a Config.

$sel:configDetails:Source', source_configDetails - Additional details for a Config, if type is dataflow endpoint or antenna demod decode.

source_dataflowSourceRegion :: Lens' Source (Maybe Text)

Region of a dataflow source.

source_configId :: Lens' Source (Maybe Text)

UUID of a Config.

source_configType :: Lens' Source (Maybe ConfigCapabilityType)

Type of a Config.

source_configDetails :: Lens' Source (Maybe ConfigDetails)

Additional details for a Config, if type is dataflow endpoint or antenna demod decode.

SpectrumConfig

data SpectrumConfig

Object that describes a spectral Config.

See: newSpectrumConfig smart constructor.

Constructors

SpectrumConfig'

Fields

polarization :: Maybe Polarization

Polarization of a spectral Config. Capturing both "RIGHT_HAND" and "LEFT_HAND" polarization requires two separate configs.

bandwidth :: FrequencyBandwidth

Bandwidth of a spectral Config. AWS Ground Station currently has the following bandwidth limitations:

For AntennaDownlinkDemodDecodeconfig, valid values are between 125 kHz to 650 MHz.

For AntennaDownlinkconfig valid values are between 10 kHz to 54 MHz.

For AntennaUplinkConfig, valid values are between 10 kHz to 54 MHz.

centerFrequency :: Frequency

Center frequency of a spectral Config. Valid values are between 2200 to 2300 MHz and 7750 to 8400 MHz for downlink and 2025 to 2120 MHz for uplink.

newSpectrumConfig

Arguments

:: FrequencyBandwidth
$sel:bandwidth:SpectrumConfig'

-> Frequency
$sel:centerFrequency:SpectrumConfig'

-> SpectrumConfig

Create a value of SpectrumConfig with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:polarization:SpectrumConfig', spectrumConfig_polarization - Polarization of a spectral Config. Capturing both "RIGHT_HAND" and "LEFT_HAND" polarization requires two separate configs.

$sel:bandwidth:SpectrumConfig', spectrumConfig_bandwidth - Bandwidth of a spectral Config. AWS Ground Station currently has the following bandwidth limitations:

For AntennaDownlinkDemodDecodeconfig, valid values are between 125 kHz to 650 MHz.

For AntennaDownlinkconfig valid values are between 10 kHz to 54 MHz.

For AntennaUplinkConfig, valid values are between 10 kHz to 54 MHz.

$sel:centerFrequency:SpectrumConfig', spectrumConfig_centerFrequency - Center frequency of a spectral Config. Valid values are between 2200 to 2300 MHz and 7750 to 8400 MHz for downlink and 2025 to 2120 MHz for uplink.

spectrumConfig_polarization :: Lens' SpectrumConfig (Maybe Polarization)

Polarization of a spectral Config. Capturing both "RIGHT_HAND" and "LEFT_HAND" polarization requires two separate configs.

spectrumConfig_bandwidth :: Lens' SpectrumConfig FrequencyBandwidth

Bandwidth of a spectral Config. AWS Ground Station currently has the following bandwidth limitations:

For AntennaDownlinkDemodDecodeconfig, valid values are between 125 kHz to 650 MHz.

For AntennaDownlinkconfig valid values are between 10 kHz to 54 MHz.

For AntennaUplinkConfig, valid values are between 10 kHz to 54 MHz.

spectrumConfig_centerFrequency :: Lens' SpectrumConfig Frequency

Center frequency of a spectral Config. Valid values are between 2200 to 2300 MHz and 7750 to 8400 MHz for downlink and 2025 to 2120 MHz for uplink.

TrackingConfig

data TrackingConfig

Object that determines whether tracking should be used during a contact executed with this Config in the mission profile.

See: newTrackingConfig smart constructor.

Constructors

TrackingConfig'

Fields

autotrack :: Criticality

Current setting for autotrack.

newTrackingConfig

Arguments

:: Criticality
$sel:autotrack:TrackingConfig'

-> TrackingConfig

Create a value of TrackingConfig with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:autotrack:TrackingConfig', trackingConfig_autotrack - Current setting for autotrack.

trackingConfig_autotrack :: Lens' TrackingConfig Criticality

Current setting for autotrack.

UplinkEchoConfig

data UplinkEchoConfig

Information about an uplink echo Config.

Parameters from the AntennaUplinkConfig, corresponding to the specified AntennaUplinkConfigArn, are used when this UplinkEchoConfig is used in a contact.

See: newUplinkEchoConfig smart constructor.

Constructors

UplinkEchoConfig'

Fields

antennaUplinkConfigArn :: Text

ARN of an uplink Config.

enabled :: Bool

Whether or not an uplink Config is enabled.

newUplinkEchoConfig

Arguments

:: Text
$sel:antennaUplinkConfigArn:UplinkEchoConfig'

-> Bool
$sel:enabled:UplinkEchoConfig'

-> UplinkEchoConfig

Create a value of UplinkEchoConfig with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:antennaUplinkConfigArn:UplinkEchoConfig', uplinkEchoConfig_antennaUplinkConfigArn - ARN of an uplink Config.

$sel:enabled:UplinkEchoConfig', uplinkEchoConfig_enabled - Whether or not an uplink Config is enabled.

uplinkEchoConfig_antennaUplinkConfigArn :: Lens' UplinkEchoConfig Text

ARN of an uplink Config.

uplinkEchoConfig_enabled :: Lens' UplinkEchoConfig Bool

Whether or not an uplink Config is enabled.

UplinkSpectrumConfig

data UplinkSpectrumConfig

Information about the uplink spectral Config.

See: newUplinkSpectrumConfig smart constructor.

Constructors

UplinkSpectrumConfig'

Fields

polarization :: Maybe Polarization

Polarization of an uplink spectral Config. Capturing both "RIGHT_HAND" and "LEFT_HAND" polarization requires two separate configs.

centerFrequency :: Frequency

Center frequency of an uplink spectral Config. Valid values are between 2025 to 2120 MHz.

newUplinkSpectrumConfig

Arguments

:: Frequency
$sel:centerFrequency:UplinkSpectrumConfig'

-> UplinkSpectrumConfig

Create a value of UplinkSpectrumConfig with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:polarization:UplinkSpectrumConfig', uplinkSpectrumConfig_polarization - Polarization of an uplink spectral Config. Capturing both "RIGHT_HAND" and "LEFT_HAND" polarization requires two separate configs.

$sel:centerFrequency:UplinkSpectrumConfig', uplinkSpectrumConfig_centerFrequency - Center frequency of an uplink spectral Config. Valid values are between 2025 to 2120 MHz.

uplinkSpectrumConfig_polarization :: Lens' UplinkSpectrumConfig (Maybe Polarization)

Polarization of an uplink spectral Config. Capturing both "RIGHT_HAND" and "LEFT_HAND" polarization requires two separate configs.

uplinkSpectrumConfig_centerFrequency :: Lens' UplinkSpectrumConfig Frequency

Center frequency of an uplink spectral Config. Valid values are between 2025 to 2120 MHz.