Amazonka.GroundStation.Types.ConfigTypeData

Description

Documentation

data ConfigTypeData

Object containing the parameters of a Config.

See the subtype definitions for what each type of Config contains.

See: newConfigTypeData smart constructor.

Constructors

ConfigTypeData'

Fields

antennaDownlinkDemodDecodeConfig :: Maybe AntennaDownlinkDemodDecodeConfig

Information about how AWS Ground Station should conﬁgure an antenna for downlink demod decode during a contact.

antennaDownlinkConfig :: Maybe AntennaDownlinkConfig

Information about how AWS Ground Station should configure an antenna for downlink during a contact.

antennaUplinkConfig :: Maybe AntennaUplinkConfig

Information about how AWS Ground Station should conﬁgure an antenna for uplink during a contact.

uplinkEchoConfig :: Maybe UplinkEchoConfig

Information about an uplink echo Config.

Parameters from the AntennaUplinkConfig, corresponding to the specified AntennaUplinkConfigArn, are used when this UplinkEchoConfig is used in a contact.

s3RecordingConfig :: Maybe S3RecordingConfig

Information about an S3 recording Config.

trackingConfig :: Maybe TrackingConfig

Object that determines whether tracking should be used during a contact executed with this Config in the mission profile.

dataflowEndpointConfig :: Maybe DataflowEndpointConfig

Information about the dataflow endpoint Config.

newConfigTypeData :: ConfigTypeData

Create a value of ConfigTypeData with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:antennaDownlinkDemodDecodeConfig:ConfigTypeData', configTypeData_antennaDownlinkDemodDecodeConfig - Information about how AWS Ground Station should conﬁgure an antenna for downlink demod decode during a contact.

$sel:antennaDownlinkConfig:ConfigTypeData', configTypeData_antennaDownlinkConfig - Information about how AWS Ground Station should configure an antenna for downlink during a contact.

$sel:antennaUplinkConfig:ConfigTypeData', configTypeData_antennaUplinkConfig - Information about how AWS Ground Station should conﬁgure an antenna for uplink during a contact.

$sel:uplinkEchoConfig:ConfigTypeData', configTypeData_uplinkEchoConfig - Information about an uplink echo Config.

Parameters from the AntennaUplinkConfig, corresponding to the specified AntennaUplinkConfigArn, are used when this UplinkEchoConfig is used in a contact.

$sel:s3RecordingConfig:ConfigTypeData', configTypeData_s3RecordingConfig - Information about an S3 recording Config.

$sel:trackingConfig:ConfigTypeData', configTypeData_trackingConfig - Object that determines whether tracking should be used during a contact executed with this Config in the mission profile.

$sel:dataflowEndpointConfig:ConfigTypeData', configTypeData_dataflowEndpointConfig - Information about the dataflow endpoint Config.

configTypeData_antennaDownlinkDemodDecodeConfig :: Lens' ConfigTypeData (Maybe AntennaDownlinkDemodDecodeConfig)

Information about how AWS Ground Station should conﬁgure an antenna for downlink demod decode during a contact.

configTypeData_antennaDownlinkConfig :: Lens' ConfigTypeData (Maybe AntennaDownlinkConfig)

Information about how AWS Ground Station should configure an antenna for downlink during a contact.

configTypeData_antennaUplinkConfig :: Lens' ConfigTypeData (Maybe AntennaUplinkConfig)

Information about how AWS Ground Station should conﬁgure an antenna for uplink during a contact.

configTypeData_uplinkEchoConfig :: Lens' ConfigTypeData (Maybe UplinkEchoConfig)

Information about an uplink echo Config.

Parameters from the AntennaUplinkConfig, corresponding to the specified AntennaUplinkConfigArn, are used when this UplinkEchoConfig is used in a contact.

configTypeData_s3RecordingConfig :: Lens' ConfigTypeData (Maybe S3RecordingConfig)

Information about an S3 recording Config.

configTypeData_trackingConfig :: Lens' ConfigTypeData (Maybe TrackingConfig)

Object that determines whether tracking should be used during a contact executed with this Config in the mission profile.

configTypeData_dataflowEndpointConfig :: Lens' ConfigTypeData (Maybe DataflowEndpointConfig)

Information about the dataflow endpoint Config.