Amazonka.GroundStation.Types.TrackingConfig

Description

Documentation

data TrackingConfig

Object that determines whether tracking should be used during a contact executed with this Config in the mission profile.

See: newTrackingConfig smart constructor.

Constructors

TrackingConfig'

Fields

autotrack :: Criticality

Current setting for autotrack.

newTrackingConfig

Arguments

:: Criticality
$sel:autotrack:TrackingConfig'

-> TrackingConfig

Create a value of TrackingConfig with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:autotrack:TrackingConfig', trackingConfig_autotrack - Current setting for autotrack.

trackingConfig_autotrack :: Lens' TrackingConfig Criticality

Current setting for autotrack.