Amazonka.GroundStation.UpdateConfig

Description

Updates the Config used when scheduling contacts.

Updating a Config will not update the execution parameters for existing future contacts scheduled with this Config.

Creating a Request

data UpdateConfig

See: newUpdateConfig smart constructor.

Constructors

UpdateConfig'

Fields

configData :: ConfigTypeData

Parameters of a Config.

configId :: Text

UUID of a Config.

configType :: ConfigCapabilityType

Type of a Config.

name :: Text

Name of a Config.

newUpdateConfig

Arguments

:: ConfigTypeData
$sel:configData:UpdateConfig'

-> Text
$sel:configId:UpdateConfig'

-> ConfigCapabilityType
$sel:configType:UpdateConfig'

-> Text
$sel:name:UpdateConfig'

-> UpdateConfig

Create a value of UpdateConfig with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:configData:UpdateConfig', updateConfig_configData - Parameters of a Config.

$sel:configId:UpdateConfig', updateConfig_configId - UUID of a Config.

$sel:configType:UpdateConfig', updateConfig_configType - Type of a Config.

$sel:name:UpdateConfig', updateConfig_name - Name of a Config.

Request Lenses

updateConfig_configData :: Lens' UpdateConfig ConfigTypeData

Parameters of a Config.

updateConfig_configId :: Lens' UpdateConfig Text

UUID of a Config.

updateConfig_configType :: Lens' UpdateConfig ConfigCapabilityType

Type of a Config.

updateConfig_name :: Lens' UpdateConfig Text

Name of a Config.

Destructuring the Response

data ConfigIdResponse

See: newConfigIdResponse smart constructor.

Constructors

ConfigIdResponse'

Fields

configArn :: Maybe Text

ARN of a Config.

configId :: Maybe Text

UUID of a Config.

configType :: Maybe ConfigCapabilityType

Type of a Config.

newConfigIdResponse :: ConfigIdResponse

Create a value of ConfigIdResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:configArn:ConfigIdResponse', configIdResponse_configArn - ARN of a Config.

$sel:configId:ConfigIdResponse', configIdResponse_configId - UUID of a Config.

$sel:configType:ConfigIdResponse', configIdResponse_configType - Type of a Config.

Response Lenses

configIdResponse_configArn :: Lens' ConfigIdResponse (Maybe Text)

ARN of a Config.

configIdResponse_configId :: Lens' ConfigIdResponse (Maybe Text)

UUID of a Config.

configIdResponse_configType :: Lens' ConfigIdResponse (Maybe ConfigCapabilityType)

Type of a Config.