Amazonka.GuardDuty.CreateDetector

Description

Creates a single Amazon GuardDuty detector. A detector is a resource that represents the GuardDuty service. To start using GuardDuty, you must create a detector in each Region where you enable the service. You can have only one detector per account per Region. All data sources are enabled in a new detector by default.

Creating a Request

data CreateDetector

See: newCreateDetector smart constructor.

Constructors

CreateDetector'

Fields

clientToken :: Maybe Text

The idempotency token for the create request.

findingPublishingFrequency :: Maybe FindingPublishingFrequency

A value that specifies how frequently updated findings are exported.

dataSources :: Maybe DataSourceConfigurations

Describes which data sources will be enabled for the detector.

tags :: Maybe (HashMap Text Text)

The tags to be added to a new detector resource.

enable :: Bool

A Boolean value that specifies whether the detector is to be enabled.

newCreateDetector

Arguments

:: Bool
$sel:enable:CreateDetector'

-> CreateDetector

Create a value of CreateDetector with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:clientToken:CreateDetector', createDetector_clientToken - The idempotency token for the create request.

$sel:findingPublishingFrequency:CreateDetector', createDetector_findingPublishingFrequency - A value that specifies how frequently updated findings are exported.

$sel:dataSources:CreateDetector', createDetector_dataSources - Describes which data sources will be enabled for the detector.

$sel:tags:CreateDetector', createDetector_tags - The tags to be added to a new detector resource.

$sel:enable:CreateDetector', createDetector_enable - A Boolean value that specifies whether the detector is to be enabled.

Request Lenses

createDetector_clientToken :: Lens' CreateDetector (Maybe Text)

The idempotency token for the create request.

createDetector_findingPublishingFrequency :: Lens' CreateDetector (Maybe FindingPublishingFrequency)

A value that specifies how frequently updated findings are exported.

createDetector_dataSources :: Lens' CreateDetector (Maybe DataSourceConfigurations)

Describes which data sources will be enabled for the detector.

createDetector_tags :: Lens' CreateDetector (Maybe (HashMap Text Text))

The tags to be added to a new detector resource.

createDetector_enable :: Lens' CreateDetector Bool

A Boolean value that specifies whether the detector is to be enabled.

Destructuring the Response

data CreateDetectorResponse

See: newCreateDetectorResponse smart constructor.

Constructors

CreateDetectorResponse'

Fields

detectorId :: Maybe Text

The unique ID of the created detector.

httpStatus :: Int

The response's http status code.

newCreateDetectorResponse

Arguments

:: Int
$sel:httpStatus:CreateDetectorResponse'

-> CreateDetectorResponse

Create a value of CreateDetectorResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:detectorId:CreateDetectorResponse', createDetectorResponse_detectorId - The unique ID of the created detector.

$sel:httpStatus:CreateDetectorResponse', createDetectorResponse_httpStatus - The response's http status code.

Response Lenses

createDetectorResponse_detectorId :: Lens' CreateDetectorResponse (Maybe Text)

The unique ID of the created detector.

createDetectorResponse_httpStatus :: Lens' CreateDetectorResponse Int

The response's http status code.