Amazonka.GuardDuty.CreateThreatIntelSet

Description

Creates a new ThreatIntelSet. ThreatIntelSets consist of known malicious IP addresses. GuardDuty generates findings based on ThreatIntelSets. Only users of the administrator account can use this operation.

Creating a Request

data CreateThreatIntelSet

See: newCreateThreatIntelSet smart constructor.

Constructors

CreateThreatIntelSet'

Fields

clientToken :: Maybe Text

The idempotency token for the create request.

tags :: Maybe (HashMap Text Text)

The tags to be added to a new threat list resource.

detectorId :: Text

The unique ID of the detector of the GuardDuty account that you want to create a threatIntelSet for.

name :: Text

A user-friendly ThreatIntelSet name displayed in all findings that are generated by activity that involves IP addresses included in this ThreatIntelSet.

format :: ThreatIntelSetFormat

The format of the file that contains the ThreatIntelSet.

location :: Text

The URI of the file that contains the ThreatIntelSet. For example: https://s3.us-west-2.amazonaws.com/my-bucket/my-object-key.

activate :: Bool

A Boolean value that indicates whether GuardDuty is to start using the uploaded ThreatIntelSet.

newCreateThreatIntelSet

Arguments

:: Text
$sel:detectorId:CreateThreatIntelSet'

-> Text
$sel:name:CreateThreatIntelSet'

-> ThreatIntelSetFormat
$sel:format:CreateThreatIntelSet'

-> Text
$sel:location:CreateThreatIntelSet'

-> Bool
$sel:activate:CreateThreatIntelSet'

-> CreateThreatIntelSet

Create a value of CreateThreatIntelSet with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:clientToken:CreateThreatIntelSet', createThreatIntelSet_clientToken - The idempotency token for the create request.

$sel:tags:CreateThreatIntelSet', createThreatIntelSet_tags - The tags to be added to a new threat list resource.

$sel:detectorId:CreateThreatIntelSet', createThreatIntelSet_detectorId - The unique ID of the detector of the GuardDuty account that you want to create a threatIntelSet for.

$sel:name:CreateThreatIntelSet', createThreatIntelSet_name - A user-friendly ThreatIntelSet name displayed in all findings that are generated by activity that involves IP addresses included in this ThreatIntelSet.

$sel:format:CreateThreatIntelSet', createThreatIntelSet_format - The format of the file that contains the ThreatIntelSet.

$sel:location:CreateThreatIntelSet', createThreatIntelSet_location - The URI of the file that contains the ThreatIntelSet. For example: https://s3.us-west-2.amazonaws.com/my-bucket/my-object-key.

$sel:activate:CreateThreatIntelSet', createThreatIntelSet_activate - A Boolean value that indicates whether GuardDuty is to start using the uploaded ThreatIntelSet.

Request Lenses

createThreatIntelSet_clientToken :: Lens' CreateThreatIntelSet (Maybe Text)

The idempotency token for the create request.

createThreatIntelSet_tags :: Lens' CreateThreatIntelSet (Maybe (HashMap Text Text))

The tags to be added to a new threat list resource.

createThreatIntelSet_detectorId :: Lens' CreateThreatIntelSet Text

The unique ID of the detector of the GuardDuty account that you want to create a threatIntelSet for.

createThreatIntelSet_name :: Lens' CreateThreatIntelSet Text

A user-friendly ThreatIntelSet name displayed in all findings that are generated by activity that involves IP addresses included in this ThreatIntelSet.

createThreatIntelSet_format :: Lens' CreateThreatIntelSet ThreatIntelSetFormat

The format of the file that contains the ThreatIntelSet.

createThreatIntelSet_location :: Lens' CreateThreatIntelSet Text

The URI of the file that contains the ThreatIntelSet. For example: https://s3.us-west-2.amazonaws.com/my-bucket/my-object-key.

createThreatIntelSet_activate :: Lens' CreateThreatIntelSet Bool

A Boolean value that indicates whether GuardDuty is to start using the uploaded ThreatIntelSet.

Destructuring the Response

data CreateThreatIntelSetResponse

See: newCreateThreatIntelSetResponse smart constructor.

Constructors

CreateThreatIntelSetResponse'

Fields

httpStatus :: Int

The response's http status code.

threatIntelSetId :: Text

The ID of the ThreatIntelSet resource.

newCreateThreatIntelSetResponse

Arguments

:: Int
$sel:httpStatus:CreateThreatIntelSetResponse'

-> Text
$sel:threatIntelSetId:CreateThreatIntelSetResponse'

-> CreateThreatIntelSetResponse

Create a value of CreateThreatIntelSetResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:httpStatus:CreateThreatIntelSetResponse', createThreatIntelSetResponse_httpStatus - The response's http status code.

$sel:threatIntelSetId:CreateThreatIntelSetResponse', createThreatIntelSetResponse_threatIntelSetId - The ID of the ThreatIntelSet resource.

Response Lenses

createThreatIntelSetResponse_httpStatus :: Lens' CreateThreatIntelSetResponse Int

The response's http status code.

createThreatIntelSetResponse_threatIntelSetId :: Lens' CreateThreatIntelSetResponse Text

The ID of the ThreatIntelSet resource.