Amazonka.GuardDuty.DeclineInvitations

Description

Declines invitations sent to the current member account by AWS accounts specified by their account IDs.

Creating a Request

data DeclineInvitations

See: newDeclineInvitations smart constructor.

Constructors

DeclineInvitations'

Fields

accountIds :: NonEmpty Text

A list of account IDs of the AWS accounts that sent invitations to the current member account that you want to decline invitations from.

newDeclineInvitations

Arguments

:: NonEmpty Text
$sel:accountIds:DeclineInvitations'

-> DeclineInvitations

Create a value of DeclineInvitations with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:accountIds:DeclineInvitations', declineInvitations_accountIds - A list of account IDs of the AWS accounts that sent invitations to the current member account that you want to decline invitations from.

Request Lenses

declineInvitations_accountIds :: Lens' DeclineInvitations (NonEmpty Text)

A list of account IDs of the AWS accounts that sent invitations to the current member account that you want to decline invitations from.

Destructuring the Response

data DeclineInvitationsResponse

See: newDeclineInvitationsResponse smart constructor.

Constructors

DeclineInvitationsResponse'

Fields

httpStatus :: Int

The response's http status code.

unprocessedAccounts :: [UnprocessedAccount]

A list of objects that contain the unprocessed account and a result string that explains why it was unprocessed.

newDeclineInvitationsResponse

Arguments

:: Int
$sel:httpStatus:DeclineInvitationsResponse'

-> DeclineInvitationsResponse

Create a value of DeclineInvitationsResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:httpStatus:DeclineInvitationsResponse', declineInvitationsResponse_httpStatus - The response's http status code.

$sel:unprocessedAccounts:DeclineInvitationsResponse', declineInvitationsResponse_unprocessedAccounts - A list of objects that contain the unprocessed account and a result string that explains why it was unprocessed.

Response Lenses

declineInvitationsResponse_httpStatus :: Lens' DeclineInvitationsResponse Int

The response's http status code.

declineInvitationsResponse_unprocessedAccounts :: Lens' DeclineInvitationsResponse [UnprocessedAccount]

A list of objects that contain the unprocessed account and a result string that explains why it was unprocessed.