Amazonka.GuardDuty.GetIPSet

Description

Retrieves the IPSet specified by the ipSetId.

Creating a Request

data GetIPSet

See: newGetIPSet smart constructor.

Constructors

GetIPSet'

Fields

detectorId :: Text

The unique ID of the detector that the IPSet is associated with.

ipSetId :: Text

The unique ID of the IPSet to retrieve.

newGetIPSet

Arguments

:: Text
$sel:detectorId:GetIPSet'

-> Text
$sel:ipSetId:GetIPSet'

-> GetIPSet

Create a value of GetIPSet with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:detectorId:GetIPSet', getIPSet_detectorId - The unique ID of the detector that the IPSet is associated with.

$sel:ipSetId:GetIPSet', getIPSet_ipSetId - The unique ID of the IPSet to retrieve.

Request Lenses

getIPSet_detectorId :: Lens' GetIPSet Text

The unique ID of the detector that the IPSet is associated with.

getIPSet_ipSetId :: Lens' GetIPSet Text

The unique ID of the IPSet to retrieve.

Destructuring the Response

data GetIPSetResponse

See: newGetIPSetResponse smart constructor.

Constructors

GetIPSetResponse'

Fields

tags :: Maybe (HashMap Text Text)

The tags of the IPSet resource.

httpStatus :: Int

The response's http status code.

name :: Text

The user-friendly name for the IPSet.

format :: IpSetFormat

The format of the file that contains the IPSet.

location :: Text

The URI of the file that contains the IPSet. For example: https://s3.us-west-2.amazonaws.com/my-bucket/my-object-key.

status :: IpSetStatus

The status of IPSet file that was uploaded.

newGetIPSetResponse

Arguments

:: Int
$sel:httpStatus:GetIPSetResponse'

-> Text
$sel:name:GetIPSetResponse'

-> IpSetFormat
$sel:format:GetIPSetResponse'

-> Text
$sel:location:GetIPSetResponse'

-> IpSetStatus
$sel:status:GetIPSetResponse'

-> GetIPSetResponse

Create a value of GetIPSetResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:tags:GetIPSetResponse', getIPSetResponse_tags - The tags of the IPSet resource.

$sel:httpStatus:GetIPSetResponse', getIPSetResponse_httpStatus - The response's http status code.

$sel:name:GetIPSetResponse', getIPSetResponse_name - The user-friendly name for the IPSet.

$sel:format:GetIPSetResponse', getIPSetResponse_format - The format of the file that contains the IPSet.

$sel:location:GetIPSetResponse', getIPSetResponse_location - The URI of the file that contains the IPSet. For example: https://s3.us-west-2.amazonaws.com/my-bucket/my-object-key.

$sel:status:GetIPSetResponse', getIPSetResponse_status - The status of IPSet file that was uploaded.

Response Lenses

getIPSetResponse_tags :: Lens' GetIPSetResponse (Maybe (HashMap Text Text))

The tags of the IPSet resource.

getIPSetResponse_httpStatus :: Lens' GetIPSetResponse Int

The response's http status code.

getIPSetResponse_name :: Lens' GetIPSetResponse Text

The user-friendly name for the IPSet.

getIPSetResponse_format :: Lens' GetIPSetResponse IpSetFormat

The format of the file that contains the IPSet.

getIPSetResponse_location :: Lens' GetIPSetResponse Text

The URI of the file that contains the IPSet. For example: https://s3.us-west-2.amazonaws.com/my-bucket/my-object-key.

getIPSetResponse_status :: Lens' GetIPSetResponse IpSetStatus

The status of IPSet file that was uploaded.