Amazonka.GuardDuty.Types.Finding

Description

Documentation

data Finding

Contains information about the finding, which is generated when abnormal or suspicious activity is detected.

See: newFinding smart constructor.

Constructors

Finding'

Fields

service :: Maybe ServiceInfo

confidence :: Maybe Double

The confidence score for the finding.

partition :: Maybe Text

The partition associated with the finding.

title :: Maybe Text

The title of the finding.

description :: Maybe Text

The description of the finding.

accountId :: Text

The ID of the account in which the finding was generated.

arn :: Text

The ARN of the finding.

createdAt :: Text

The time and date when the finding was created.

id :: Text

The ID of the finding.

region :: Text

The Region where the finding was generated.

resource :: Resource

schemaVersion :: Text

The version of the schema used for the finding.

severity :: Double

The severity of the finding.

type' :: Text

The type of finding.

updatedAt :: Text

The time and date when the finding was last updated.

newFinding

Arguments

:: Text
$sel:accountId:Finding'

-> Text
$sel:arn:Finding'

-> Text
$sel:createdAt:Finding'

-> Text
$sel:id:Finding'

-> Text
$sel:region:Finding'

-> Resource
$sel:resource:Finding'

-> Text
$sel:schemaVersion:Finding'

-> Double
$sel:severity:Finding'

-> Text
$sel:type':Finding'

-> Text
$sel:updatedAt:Finding'

-> Finding

Create a value of Finding with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:service:Finding', finding_service - Undocumented member.

$sel:confidence:Finding', finding_confidence - The confidence score for the finding.

$sel:partition:Finding', finding_partition - The partition associated with the finding.

$sel:title:Finding', finding_title - The title of the finding.

$sel:description:Finding', finding_description - The description of the finding.

$sel:accountId:Finding', finding_accountId - The ID of the account in which the finding was generated.

$sel:arn:Finding', finding_arn - The ARN of the finding.

$sel:createdAt:Finding', finding_createdAt - The time and date when the finding was created.

$sel:id:Finding', finding_id - The ID of the finding.

$sel:region:Finding', finding_region - The Region where the finding was generated.

$sel:resource:Finding', finding_resource - Undocumented member.

$sel:schemaVersion:Finding', finding_schemaVersion - The version of the schema used for the finding.

$sel:severity:Finding', finding_severity - The severity of the finding.

$sel:type':Finding', finding_type - The type of finding.

$sel:updatedAt:Finding', finding_updatedAt - The time and date when the finding was last updated.

finding_service :: Lens' Finding (Maybe ServiceInfo)

Undocumented member.

finding_confidence :: Lens' Finding (Maybe Double)

The confidence score for the finding.

finding_partition :: Lens' Finding (Maybe Text)

The partition associated with the finding.

finding_title :: Lens' Finding (Maybe Text)

The title of the finding.

finding_description :: Lens' Finding (Maybe Text)

The description of the finding.

finding_accountId :: Lens' Finding Text

The ID of the account in which the finding was generated.

finding_arn :: Lens' Finding Text

The ARN of the finding.

finding_createdAt :: Lens' Finding Text

The time and date when the finding was created.

finding_id :: Lens' Finding Text

The ID of the finding.

finding_region :: Lens' Finding Text

The Region where the finding was generated.

finding_resource :: Lens' Finding Resource

Undocumented member.

finding_schemaVersion :: Lens' Finding Text

The version of the schema used for the finding.

finding_severity :: Lens' Finding Double

The severity of the finding.

finding_type :: Lens' Finding Text

The type of finding.

finding_updatedAt :: Lens' Finding Text

The time and date when the finding was last updated.