Amazonka.GuardDuty.Types.Member

Description

Documentation

data Member

Contains information about the member account.

See: newMember smart constructor.

Constructors

Member'

Fields

invitedAt :: Maybe Text

The timestamp when the invitation was sent.

detectorId :: Maybe Text

The detector ID of the member account.

accountId :: Text

The ID of the member account.

masterId :: Text

The administrator account ID.

email :: Text

The email address of the member account.

relationshipStatus :: Text

The status of the relationship between the member and the administrator.

updatedAt :: Text

The last-updated timestamp of the member.

newMember

Arguments

:: Text
$sel:accountId:Member'

-> Text
$sel:masterId:Member'

-> Text
$sel:email:Member'

-> Text
$sel:relationshipStatus:Member'

-> Text
$sel:updatedAt:Member'

-> Member

Create a value of Member with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:invitedAt:Member', member_invitedAt - The timestamp when the invitation was sent.

$sel:detectorId:Member', member_detectorId - The detector ID of the member account.

$sel:accountId:Member', member_accountId - The ID of the member account.

$sel:masterId:Member', member_masterId - The administrator account ID.

$sel:email:Member', member_email - The email address of the member account.

$sel:relationshipStatus:Member', member_relationshipStatus - The status of the relationship between the member and the administrator.

$sel:updatedAt:Member', member_updatedAt - The last-updated timestamp of the member.

member_invitedAt :: Lens' Member (Maybe Text)

The timestamp when the invitation was sent.

member_detectorId :: Lens' Member (Maybe Text)

The detector ID of the member account.

member_accountId :: Lens' Member Text

The ID of the member account.

member_masterId :: Lens' Member Text

The administrator account ID.

member_email :: Lens' Member Text

The email address of the member account.

member_relationshipStatus :: Lens' Member Text

The status of the relationship between the member and the administrator.

member_updatedAt :: Lens' Member Text

The last-updated timestamp of the member.