Amazonka.GuardDuty.UpdateIPSet

Description

Updates the IPSet specified by the IPSet ID.

Creating a Request

data UpdateIPSet

See: newUpdateIPSet smart constructor.

Constructors

UpdateIPSet'

Fields

location :: Maybe Text

The updated URI of the file that contains the IPSet. For example: https://s3.us-west-2.amazonaws.com/my-bucket/my-object-key.

activate :: Maybe Bool

The updated Boolean value that specifies whether the IPSet is active or not.

name :: Maybe Text

The unique ID that specifies the IPSet that you want to update.

detectorId :: Text

The detectorID that specifies the GuardDuty service whose IPSet you want to update.

ipSetId :: Text

The unique ID that specifies the IPSet that you want to update.

newUpdateIPSet

Arguments

:: Text
$sel:detectorId:UpdateIPSet'

-> Text
$sel:ipSetId:UpdateIPSet'

-> UpdateIPSet

Create a value of UpdateIPSet with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:location:UpdateIPSet', updateIPSet_location - The updated URI of the file that contains the IPSet. For example: https://s3.us-west-2.amazonaws.com/my-bucket/my-object-key.

$sel:activate:UpdateIPSet', updateIPSet_activate - The updated Boolean value that specifies whether the IPSet is active or not.

$sel:name:UpdateIPSet', updateIPSet_name - The unique ID that specifies the IPSet that you want to update.

$sel:detectorId:UpdateIPSet', updateIPSet_detectorId - The detectorID that specifies the GuardDuty service whose IPSet you want to update.

$sel:ipSetId:UpdateIPSet', updateIPSet_ipSetId - The unique ID that specifies the IPSet that you want to update.

Request Lenses

updateIPSet_location :: Lens' UpdateIPSet (Maybe Text)

The updated URI of the file that contains the IPSet. For example: https://s3.us-west-2.amazonaws.com/my-bucket/my-object-key.

updateIPSet_activate :: Lens' UpdateIPSet (Maybe Bool)

The updated Boolean value that specifies whether the IPSet is active or not.

updateIPSet_name :: Lens' UpdateIPSet (Maybe Text)

The unique ID that specifies the IPSet that you want to update.

updateIPSet_detectorId :: Lens' UpdateIPSet Text

The detectorID that specifies the GuardDuty service whose IPSet you want to update.

updateIPSet_ipSetId :: Lens' UpdateIPSet Text

The unique ID that specifies the IPSet that you want to update.

Destructuring the Response

data UpdateIPSetResponse

See: newUpdateIPSetResponse smart constructor.

Constructors

UpdateIPSetResponse'

Fields

httpStatus :: Int

The response's http status code.

newUpdateIPSetResponse

Arguments

:: Int
$sel:httpStatus:UpdateIPSetResponse'

-> UpdateIPSetResponse

Create a value of UpdateIPSetResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:httpStatus:UpdateIPSetResponse', updateIPSetResponse_httpStatus - The response's http status code.

Response Lenses

updateIPSetResponse_httpStatus :: Lens' UpdateIPSetResponse Int

The response's http status code.