Amazonka.HealthLake.CreateFHIRDatastore

Description

Creates a Data Store that can ingest and export FHIR formatted data.

Creating a Request

data CreateFHIRDatastore

See: newCreateFHIRDatastore smart constructor.

Constructors

CreateFHIRDatastore'

Fields

clientToken :: Maybe Text

Optional user provided token used for ensuring idempotency.

sseConfiguration :: Maybe SseConfiguration

The server-side encryption key configuration for a customer provided encryption key specified for creating a Data Store.

datastoreName :: Maybe Text

The user generated name for the Data Store.

preloadDataConfig :: Maybe PreloadDataConfig

Optional parameter to preload data upon creation of the Data Store. Currently, the only supported preloaded data is synthetic data generated from Synthea.

tags :: Maybe [Tag]

Resource tags that are applied to a Data Store when it is created.

datastoreTypeVersion :: FHIRVersion

The FHIR version of the Data Store. The only supported version is R4.

newCreateFHIRDatastore

Arguments

:: FHIRVersion
$sel:datastoreTypeVersion:CreateFHIRDatastore'

-> CreateFHIRDatastore

Create a value of CreateFHIRDatastore with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:clientToken:CreateFHIRDatastore', createFHIRDatastore_clientToken - Optional user provided token used for ensuring idempotency.

$sel:sseConfiguration:CreateFHIRDatastore', createFHIRDatastore_sseConfiguration - The server-side encryption key configuration for a customer provided encryption key specified for creating a Data Store.

$sel:datastoreName:CreateFHIRDatastore', createFHIRDatastore_datastoreName - The user generated name for the Data Store.

$sel:preloadDataConfig:CreateFHIRDatastore', createFHIRDatastore_preloadDataConfig - Optional parameter to preload data upon creation of the Data Store. Currently, the only supported preloaded data is synthetic data generated from Synthea.

$sel:tags:CreateFHIRDatastore', createFHIRDatastore_tags - Resource tags that are applied to a Data Store when it is created.

$sel:datastoreTypeVersion:CreateFHIRDatastore', createFHIRDatastore_datastoreTypeVersion - The FHIR version of the Data Store. The only supported version is R4.

Request Lenses

createFHIRDatastore_clientToken :: Lens' CreateFHIRDatastore (Maybe Text)

Optional user provided token used for ensuring idempotency.

createFHIRDatastore_sseConfiguration :: Lens' CreateFHIRDatastore (Maybe SseConfiguration)

The server-side encryption key configuration for a customer provided encryption key specified for creating a Data Store.

createFHIRDatastore_datastoreName :: Lens' CreateFHIRDatastore (Maybe Text)

The user generated name for the Data Store.

createFHIRDatastore_preloadDataConfig :: Lens' CreateFHIRDatastore (Maybe PreloadDataConfig)

Optional parameter to preload data upon creation of the Data Store. Currently, the only supported preloaded data is synthetic data generated from Synthea.

createFHIRDatastore_tags :: Lens' CreateFHIRDatastore (Maybe [Tag])

Resource tags that are applied to a Data Store when it is created.

createFHIRDatastore_datastoreTypeVersion :: Lens' CreateFHIRDatastore FHIRVersion

The FHIR version of the Data Store. The only supported version is R4.

Destructuring the Response

data CreateFHIRDatastoreResponse

See: newCreateFHIRDatastoreResponse smart constructor.

Constructors

CreateFHIRDatastoreResponse'

Fields

httpStatus :: Int

The response's http status code.

datastoreId :: Text

The AWS-generated Data Store id. This id is in the output from the initial Data Store creation call.

datastoreArn :: Text

The datastore ARN is generated during the creation of the Data Store and can be found in the output from the initial Data Store creation call.

datastoreStatus :: DatastoreStatus

The status of the FHIR Data Store. Possible statuses are ‘CREATING’, ‘ACTIVE’, ‘DELETING’, ‘DELETED’.

datastoreEndpoint :: Text

The AWS endpoint for the created Data Store. For preview, only US-east-1 endpoints are supported.

newCreateFHIRDatastoreResponse

Arguments

:: Int
$sel:httpStatus:CreateFHIRDatastoreResponse'

-> Text
$sel:datastoreId:CreateFHIRDatastoreResponse'

-> Text
$sel:datastoreArn:CreateFHIRDatastoreResponse'

-> DatastoreStatus
$sel:datastoreStatus:CreateFHIRDatastoreResponse'

-> Text
$sel:datastoreEndpoint:CreateFHIRDatastoreResponse'

-> CreateFHIRDatastoreResponse

Create a value of CreateFHIRDatastoreResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:httpStatus:CreateFHIRDatastoreResponse', createFHIRDatastoreResponse_httpStatus - The response's http status code.

$sel:datastoreId:CreateFHIRDatastoreResponse', createFHIRDatastoreResponse_datastoreId - The AWS-generated Data Store id. This id is in the output from the initial Data Store creation call.

$sel:datastoreArn:CreateFHIRDatastoreResponse', createFHIRDatastoreResponse_datastoreArn - The datastore ARN is generated during the creation of the Data Store and can be found in the output from the initial Data Store creation call.

$sel:datastoreStatus:CreateFHIRDatastoreResponse', createFHIRDatastoreResponse_datastoreStatus - The status of the FHIR Data Store. Possible statuses are ‘CREATING’, ‘ACTIVE’, ‘DELETING’, ‘DELETED’.

$sel:datastoreEndpoint:CreateFHIRDatastoreResponse', createFHIRDatastoreResponse_datastoreEndpoint - The AWS endpoint for the created Data Store. For preview, only US-east-1 endpoints are supported.

Response Lenses

createFHIRDatastoreResponse_httpStatus :: Lens' CreateFHIRDatastoreResponse Int

The response's http status code.

createFHIRDatastoreResponse_datastoreId :: Lens' CreateFHIRDatastoreResponse Text

The AWS-generated Data Store id. This id is in the output from the initial Data Store creation call.

createFHIRDatastoreResponse_datastoreArn :: Lens' CreateFHIRDatastoreResponse Text

The datastore ARN is generated during the creation of the Data Store and can be found in the output from the initial Data Store creation call.

createFHIRDatastoreResponse_datastoreStatus :: Lens' CreateFHIRDatastoreResponse DatastoreStatus

The status of the FHIR Data Store. Possible statuses are ‘CREATING’, ‘ACTIVE’, ‘DELETING’, ‘DELETED’.

createFHIRDatastoreResponse_datastoreEndpoint :: Lens' CreateFHIRDatastoreResponse Text

The AWS endpoint for the created Data Store. For preview, only US-east-1 endpoints are supported.