Amazonka.HealthLake.DescribeFHIRImportJob

Description

Displays the properties of a FHIR import job, including the ID, ARN, name, and the status of the job.

Creating a Request

data DescribeFHIRImportJob

See: newDescribeFHIRImportJob smart constructor.

Constructors

DescribeFHIRImportJob'

Fields

datastoreId :: Text

The AWS-generated ID of the Data Store.

jobId :: Text

The AWS-generated job ID.

newDescribeFHIRImportJob

Arguments

:: Text
$sel:datastoreId:DescribeFHIRImportJob'

-> Text
$sel:jobId:DescribeFHIRImportJob'

-> DescribeFHIRImportJob

Create a value of DescribeFHIRImportJob with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:datastoreId:DescribeFHIRImportJob', describeFHIRImportJob_datastoreId - The AWS-generated ID of the Data Store.

$sel:jobId:DescribeFHIRImportJob', describeFHIRImportJob_jobId - The AWS-generated job ID.

Request Lenses

describeFHIRImportJob_datastoreId :: Lens' DescribeFHIRImportJob Text

The AWS-generated ID of the Data Store.

describeFHIRImportJob_jobId :: Lens' DescribeFHIRImportJob Text

The AWS-generated job ID.

Destructuring the Response

data DescribeFHIRImportJobResponse

See: newDescribeFHIRImportJobResponse smart constructor.

Constructors

DescribeFHIRImportJobResponse'

Fields

httpStatus :: Int

The response's http status code.

importJobProperties :: ImportJobProperties

The properties of the Import job request, including the ID, ARN, name, and the status of the job.

newDescribeFHIRImportJobResponse

Arguments

:: Int
$sel:httpStatus:DescribeFHIRImportJobResponse'

-> ImportJobProperties
$sel:importJobProperties:DescribeFHIRImportJobResponse'

-> DescribeFHIRImportJobResponse

Create a value of DescribeFHIRImportJobResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:httpStatus:DescribeFHIRImportJobResponse', describeFHIRImportJobResponse_httpStatus - The response's http status code.

$sel:importJobProperties:DescribeFHIRImportJobResponse', describeFHIRImportJobResponse_importJobProperties - The properties of the Import job request, including the ID, ARN, name, and the status of the job.

Response Lenses

describeFHIRImportJobResponse_httpStatus :: Lens' DescribeFHIRImportJobResponse Int

The response's http status code.

describeFHIRImportJobResponse_importJobProperties :: Lens' DescribeFHIRImportJobResponse ImportJobProperties

The properties of the Import job request, including the ID, ARN, name, and the status of the job.