Amazonka.HealthLake.ListFHIRImportJobs

Description

Lists all FHIR import jobs associated with an account and their statuses.

Creating a Request

data ListFHIRImportJobs

See: newListFHIRImportJobs smart constructor.

Constructors

ListFHIRImportJobs'

Fields

submittedAfter :: Maybe POSIX

This parameter limits the response to FHIR import jobs submitted after a user specified date.

jobName :: Maybe Text

This parameter limits the response to the import job with the specified job name.

submittedBefore :: Maybe POSIX

This parameter limits the response to FHIR import jobs submitted before a user specified date.

nextToken :: Maybe Text

A pagination token used to identify the next page of results to return for a ListFHIRImportJobs query.

jobStatus :: Maybe JobStatus

This parameter limits the response to the import job with the specified job status.

maxResults :: Maybe Natural

This parameter limits the number of results returned for a ListFHIRImportJobs to a maximum quantity specified by the user.

datastoreId :: Text

This parameter limits the response to the import job with the specified Data Store ID.

newListFHIRImportJobs

Arguments

:: Text
$sel:datastoreId:ListFHIRImportJobs'

-> ListFHIRImportJobs

Create a value of ListFHIRImportJobs with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:submittedAfter:ListFHIRImportJobs', listFHIRImportJobs_submittedAfter - This parameter limits the response to FHIR import jobs submitted after a user specified date.

$sel:jobName:ListFHIRImportJobs', listFHIRImportJobs_jobName - This parameter limits the response to the import job with the specified job name.

$sel:submittedBefore:ListFHIRImportJobs', listFHIRImportJobs_submittedBefore - This parameter limits the response to FHIR import jobs submitted before a user specified date.

$sel:nextToken:ListFHIRImportJobs', listFHIRImportJobs_nextToken - A pagination token used to identify the next page of results to return for a ListFHIRImportJobs query.

$sel:jobStatus:ListFHIRImportJobs', listFHIRImportJobs_jobStatus - This parameter limits the response to the import job with the specified job status.

$sel:maxResults:ListFHIRImportJobs', listFHIRImportJobs_maxResults - This parameter limits the number of results returned for a ListFHIRImportJobs to a maximum quantity specified by the user.

$sel:datastoreId:ListFHIRImportJobs', listFHIRImportJobs_datastoreId - This parameter limits the response to the import job with the specified Data Store ID.

Request Lenses

listFHIRImportJobs_submittedAfter :: Lens' ListFHIRImportJobs (Maybe UTCTime)

This parameter limits the response to FHIR import jobs submitted after a user specified date.

listFHIRImportJobs_jobName :: Lens' ListFHIRImportJobs (Maybe Text)

This parameter limits the response to the import job with the specified job name.

listFHIRImportJobs_submittedBefore :: Lens' ListFHIRImportJobs (Maybe UTCTime)

This parameter limits the response to FHIR import jobs submitted before a user specified date.

listFHIRImportJobs_nextToken :: Lens' ListFHIRImportJobs (Maybe Text)

A pagination token used to identify the next page of results to return for a ListFHIRImportJobs query.

listFHIRImportJobs_jobStatus :: Lens' ListFHIRImportJobs (Maybe JobStatus)

This parameter limits the response to the import job with the specified job status.

listFHIRImportJobs_maxResults :: Lens' ListFHIRImportJobs (Maybe Natural)

This parameter limits the number of results returned for a ListFHIRImportJobs to a maximum quantity specified by the user.

listFHIRImportJobs_datastoreId :: Lens' ListFHIRImportJobs Text

This parameter limits the response to the import job with the specified Data Store ID.

Destructuring the Response

data ListFHIRImportJobsResponse

See: newListFHIRImportJobsResponse smart constructor.

Constructors

ListFHIRImportJobsResponse'

Fields

nextToken :: Maybe Text

A pagination token used to identify the next page of results to return for a ListFHIRImportJobs query.

httpStatus :: Int

The response's http status code.

importJobPropertiesList :: [ImportJobProperties]

The properties of a listed FHIR import jobs, including the ID, ARN, name, and the status of the job.

newListFHIRImportJobsResponse

Arguments

:: Int
$sel:httpStatus:ListFHIRImportJobsResponse'

-> ListFHIRImportJobsResponse

Create a value of ListFHIRImportJobsResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:nextToken:ListFHIRImportJobs', listFHIRImportJobsResponse_nextToken - A pagination token used to identify the next page of results to return for a ListFHIRImportJobs query.

$sel:httpStatus:ListFHIRImportJobsResponse', listFHIRImportJobsResponse_httpStatus - The response's http status code.

$sel:importJobPropertiesList:ListFHIRImportJobsResponse', listFHIRImportJobsResponse_importJobPropertiesList - The properties of a listed FHIR import jobs, including the ID, ARN, name, and the status of the job.

Response Lenses

listFHIRImportJobsResponse_nextToken :: Lens' ListFHIRImportJobsResponse (Maybe Text)

A pagination token used to identify the next page of results to return for a ListFHIRImportJobs query.

listFHIRImportJobsResponse_httpStatus :: Lens' ListFHIRImportJobsResponse Int

The response's http status code.

listFHIRImportJobsResponse_importJobPropertiesList :: Lens' ListFHIRImportJobsResponse [ImportJobProperties]

The properties of a listed FHIR import jobs, including the ID, ARN, name, and the status of the job.