Amazonka.HealthLake.StartFHIRExportJob

Description

Begins a FHIR export job.

Creating a Request

data StartFHIRExportJob

See: newStartFHIRExportJob smart constructor.

Constructors

StartFHIRExportJob'

Fields

jobName :: Maybe Text

The user generated name for an export job.

outputDataConfig :: OutputDataConfig

The output data configuration that was supplied when the export job was created.

datastoreId :: Text

The AWS generated ID for the Data Store from which files are being exported for an export job.

dataAccessRoleArn :: Text

The Amazon Resource Name used during the initiation of the job.

clientToken :: Text

An optional user provided token used for ensuring idempotency.

newStartFHIRExportJob

Arguments

:: OutputDataConfig
$sel:outputDataConfig:StartFHIRExportJob'

-> Text
$sel:datastoreId:StartFHIRExportJob'

-> Text
$sel:dataAccessRoleArn:StartFHIRExportJob'

-> Text
$sel:clientToken:StartFHIRExportJob'

-> StartFHIRExportJob

Create a value of StartFHIRExportJob with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:jobName:StartFHIRExportJob', startFHIRExportJob_jobName - The user generated name for an export job.

$sel:outputDataConfig:StartFHIRExportJob', startFHIRExportJob_outputDataConfig - The output data configuration that was supplied when the export job was created.

$sel:datastoreId:StartFHIRExportJob', startFHIRExportJob_datastoreId - The AWS generated ID for the Data Store from which files are being exported for an export job.

$sel:dataAccessRoleArn:StartFHIRExportJob', startFHIRExportJob_dataAccessRoleArn - The Amazon Resource Name used during the initiation of the job.

$sel:clientToken:StartFHIRExportJob', startFHIRExportJob_clientToken - An optional user provided token used for ensuring idempotency.

Request Lenses

startFHIRExportJob_jobName :: Lens' StartFHIRExportJob (Maybe Text)

The user generated name for an export job.

startFHIRExportJob_outputDataConfig :: Lens' StartFHIRExportJob OutputDataConfig

The output data configuration that was supplied when the export job was created.

startFHIRExportJob_datastoreId :: Lens' StartFHIRExportJob Text

The AWS generated ID for the Data Store from which files are being exported for an export job.

startFHIRExportJob_dataAccessRoleArn :: Lens' StartFHIRExportJob Text

The Amazon Resource Name used during the initiation of the job.

startFHIRExportJob_clientToken :: Lens' StartFHIRExportJob Text

An optional user provided token used for ensuring idempotency.

Destructuring the Response

data StartFHIRExportJobResponse

See: newStartFHIRExportJobResponse smart constructor.

Constructors

StartFHIRExportJobResponse'

Fields

datastoreId :: Maybe Text

The AWS generated ID for the Data Store from which files are being exported for an export job.

httpStatus :: Int

The response's http status code.

jobId :: Text

The AWS generated ID for an export job.

jobStatus :: JobStatus

The status of a FHIR export job. Possible statuses are SUBMITTED, IN_PROGRESS, COMPLETED, or FAILED.

newStartFHIRExportJobResponse

Arguments

:: Int
$sel:httpStatus:StartFHIRExportJobResponse'

-> Text
$sel:jobId:StartFHIRExportJobResponse'

-> JobStatus
$sel:jobStatus:StartFHIRExportJobResponse'

-> StartFHIRExportJobResponse

Create a value of StartFHIRExportJobResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:datastoreId:StartFHIRExportJob', startFHIRExportJobResponse_datastoreId - The AWS generated ID for the Data Store from which files are being exported for an export job.

$sel:httpStatus:StartFHIRExportJobResponse', startFHIRExportJobResponse_httpStatus - The response's http status code.

$sel:jobId:StartFHIRExportJobResponse', startFHIRExportJobResponse_jobId - The AWS generated ID for an export job.

$sel:jobStatus:StartFHIRExportJobResponse', startFHIRExportJobResponse_jobStatus - The status of a FHIR export job. Possible statuses are SUBMITTED, IN_PROGRESS, COMPLETED, or FAILED.

Response Lenses

startFHIRExportJobResponse_datastoreId :: Lens' StartFHIRExportJobResponse (Maybe Text)

The AWS generated ID for the Data Store from which files are being exported for an export job.

startFHIRExportJobResponse_httpStatus :: Lens' StartFHIRExportJobResponse Int

The response's http status code.

startFHIRExportJobResponse_jobId :: Lens' StartFHIRExportJobResponse Text

The AWS generated ID for an export job.

startFHIRExportJobResponse_jobStatus :: Lens' StartFHIRExportJobResponse JobStatus

The status of a FHIR export job. Possible statuses are SUBMITTED, IN_PROGRESS, COMPLETED, or FAILED.