Amazonka.HealthLake.Types

Description

Service Configuration

defaultService :: Service

API version 2017-07-01 of the Amazon HealthLake SDK configuration.

Errors

_ValidationException :: AsError a => Getting (First ServiceError) a ServiceError

The user input parameter was invalid.

_AccessDeniedException :: AsError a => Getting (First ServiceError) a ServiceError

Access is denied. Your account is not authorized to perform this operation.

_ConflictException :: AsError a => Getting (First ServiceError) a ServiceError

The Data Store is in a transition state and the user requested action can not be performed.

_ThrottlingException :: AsError a => Getting (First ServiceError) a ServiceError

The user has exceeded their maximum number of allowed calls to the given API.

_InternalServerException :: AsError a => Getting (First ServiceError) a ServiceError

Unknown error occurs in the service.

_ResourceNotFoundException :: AsError a => Getting (First ServiceError) a ServiceError

The requested Data Store was not found.

CmkType

newtype CmkType

Constructors

CmkType'

Fields

fromCmkType :: Text

Bundled Patterns

pattern CmkType_AWS_OWNED_KMS_KEY :: CmkType

pattern CmkType_CUSTOMER_MANAGED_KMS_KEY :: CmkType

DatastoreStatus

newtype DatastoreStatus

Constructors

DatastoreStatus'

Fields

fromDatastoreStatus :: Text

Bundled Patterns

pattern DatastoreStatus_ACTIVE :: DatastoreStatus

pattern DatastoreStatus_CREATING :: DatastoreStatus

pattern DatastoreStatus_DELETED :: DatastoreStatus

pattern DatastoreStatus_DELETING :: DatastoreStatus

FHIRVersion

newtype FHIRVersion

Constructors

FHIRVersion'

Fields

fromFHIRVersion :: Text

Bundled Patterns

pattern FHIRVersion_R4 :: FHIRVersion

JobStatus

newtype JobStatus

Constructors

JobStatus'

Fields

fromJobStatus :: Text

Bundled Patterns

pattern JobStatus_COMPLETED :: JobStatus

pattern JobStatus_COMPLETED_WITH_ERRORS :: JobStatus

pattern JobStatus_FAILED :: JobStatus

pattern JobStatus_IN_PROGRESS :: JobStatus

pattern JobStatus_SUBMITTED :: JobStatus

PreloadDataType

newtype PreloadDataType

Constructors

PreloadDataType'

Fields

fromPreloadDataType :: Text

Bundled Patterns

pattern PreloadDataType_SYNTHEA :: PreloadDataType

DatastoreFilter

data DatastoreFilter

The filters applied to Data Store query.

See: newDatastoreFilter smart constructor.

Constructors

DatastoreFilter'

Fields

createdAfter :: Maybe POSIX

A filter that allows the user to set cutoff dates for records. All Data Stores created after the specified date will be included in the results.

datastoreName :: Maybe Text

Allows the user to filter Data Store results by name.

datastoreStatus :: Maybe DatastoreStatus

Allows the user to filter Data Store results by status.

createdBefore :: Maybe POSIX

A filter that allows the user to set cutoff dates for records. All Data Stores created before the specified date will be included in the results.

newDatastoreFilter :: DatastoreFilter

Create a value of DatastoreFilter with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:createdAfter:DatastoreFilter', datastoreFilter_createdAfter - A filter that allows the user to set cutoff dates for records. All Data Stores created after the specified date will be included in the results.

$sel:datastoreName:DatastoreFilter', datastoreFilter_datastoreName - Allows the user to filter Data Store results by name.

$sel:datastoreStatus:DatastoreFilter', datastoreFilter_datastoreStatus - Allows the user to filter Data Store results by status.

$sel:createdBefore:DatastoreFilter', datastoreFilter_createdBefore - A filter that allows the user to set cutoff dates for records. All Data Stores created before the specified date will be included in the results.

datastoreFilter_createdAfter :: Lens' DatastoreFilter (Maybe UTCTime)

A filter that allows the user to set cutoff dates for records. All Data Stores created after the specified date will be included in the results.

datastoreFilter_datastoreName :: Lens' DatastoreFilter (Maybe Text)

Allows the user to filter Data Store results by name.

datastoreFilter_datastoreStatus :: Lens' DatastoreFilter (Maybe DatastoreStatus)

Allows the user to filter Data Store results by status.

datastoreFilter_createdBefore :: Lens' DatastoreFilter (Maybe UTCTime)

A filter that allows the user to set cutoff dates for records. All Data Stores created before the specified date will be included in the results.

DatastoreProperties

data DatastoreProperties

Displays the properties of the Data Store, including the ID, Arn, name, and the status of the Data Store.

See: newDatastoreProperties smart constructor.

Constructors

DatastoreProperties'

Fields

sseConfiguration :: Maybe SseConfiguration

The server-side encryption key configuration for a customer provided encryption key (CMK).

createdAt :: Maybe POSIX

The time that a Data Store was created.

datastoreName :: Maybe Text

The user-generated name for the Data Store.

preloadDataConfig :: Maybe PreloadDataConfig

The preloaded data configuration for the Data Store. Only data preloaded from Synthea is supported.

datastoreId :: Text

The AWS-generated ID number for the Data Store.

datastoreArn :: Text

The Amazon Resource Name used in the creation of the Data Store.

datastoreStatus :: DatastoreStatus

The status of the Data Store. Possible statuses are 'CREATING', 'ACTIVE', 'DELETING', or 'DELETED'.

datastoreTypeVersion :: FHIRVersion

The FHIR version. Only R4 version data is supported.

datastoreEndpoint :: Text

The AWS endpoint for the Data Store. Each Data Store will have it's own endpoint with Data Store ID in the endpoint URL.

newDatastoreProperties

Arguments

:: Text
$sel:datastoreId:DatastoreProperties'

-> Text
$sel:datastoreArn:DatastoreProperties'

-> DatastoreStatus
$sel:datastoreStatus:DatastoreProperties'

-> FHIRVersion
$sel:datastoreTypeVersion:DatastoreProperties'

-> Text
$sel:datastoreEndpoint:DatastoreProperties'

-> DatastoreProperties

Create a value of DatastoreProperties with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:sseConfiguration:DatastoreProperties', datastoreProperties_sseConfiguration - The server-side encryption key configuration for a customer provided encryption key (CMK).

$sel:createdAt:DatastoreProperties', datastoreProperties_createdAt - The time that a Data Store was created.

$sel:datastoreName:DatastoreProperties', datastoreProperties_datastoreName - The user-generated name for the Data Store.

$sel:preloadDataConfig:DatastoreProperties', datastoreProperties_preloadDataConfig - The preloaded data configuration for the Data Store. Only data preloaded from Synthea is supported.

$sel:datastoreId:DatastoreProperties', datastoreProperties_datastoreId - The AWS-generated ID number for the Data Store.

$sel:datastoreArn:DatastoreProperties', datastoreProperties_datastoreArn - The Amazon Resource Name used in the creation of the Data Store.

$sel:datastoreStatus:DatastoreProperties', datastoreProperties_datastoreStatus - The status of the Data Store. Possible statuses are 'CREATING', 'ACTIVE', 'DELETING', or 'DELETED'.

$sel:datastoreTypeVersion:DatastoreProperties', datastoreProperties_datastoreTypeVersion - The FHIR version. Only R4 version data is supported.

$sel:datastoreEndpoint:DatastoreProperties', datastoreProperties_datastoreEndpoint - The AWS endpoint for the Data Store. Each Data Store will have it's own endpoint with Data Store ID in the endpoint URL.

datastoreProperties_sseConfiguration :: Lens' DatastoreProperties (Maybe SseConfiguration)

The server-side encryption key configuration for a customer provided encryption key (CMK).

datastoreProperties_createdAt :: Lens' DatastoreProperties (Maybe UTCTime)

The time that a Data Store was created.

datastoreProperties_datastoreName :: Lens' DatastoreProperties (Maybe Text)

The user-generated name for the Data Store.

datastoreProperties_preloadDataConfig :: Lens' DatastoreProperties (Maybe PreloadDataConfig)

The preloaded data configuration for the Data Store. Only data preloaded from Synthea is supported.

datastoreProperties_datastoreId :: Lens' DatastoreProperties Text

The AWS-generated ID number for the Data Store.

datastoreProperties_datastoreArn :: Lens' DatastoreProperties Text

The Amazon Resource Name used in the creation of the Data Store.

datastoreProperties_datastoreStatus :: Lens' DatastoreProperties DatastoreStatus

The status of the Data Store. Possible statuses are 'CREATING', 'ACTIVE', 'DELETING', or 'DELETED'.

datastoreProperties_datastoreTypeVersion :: Lens' DatastoreProperties FHIRVersion

The FHIR version. Only R4 version data is supported.

datastoreProperties_datastoreEndpoint :: Lens' DatastoreProperties Text

The AWS endpoint for the Data Store. Each Data Store will have it's own endpoint with Data Store ID in the endpoint URL.

ExportJobProperties

data ExportJobProperties

The properties of a FHIR export job, including the ID, ARN, name, and the status of the job.

See: newExportJobProperties smart constructor.

Constructors

ExportJobProperties'

Fields

jobName :: Maybe Text

The user generated name for an export job.

endTime :: Maybe POSIX

The time an export job completed.

dataAccessRoleArn :: Maybe Text

The Amazon Resource Name used during the initiation of the job.

message :: Maybe Text

An explanation of any errors that may have occurred during the export job.

jobId :: Text

The AWS generated ID for an export job.

jobStatus :: JobStatus

The status of a FHIR export job. Possible statuses are SUBMITTED, IN_PROGRESS, COMPLETED, or FAILED.

submitTime :: POSIX

The time an export job was initiated.

datastoreId :: Text

The AWS generated ID for the Data Store from which files are being exported for an export job.

outputDataConfig :: OutputDataConfig

The output data configuration that was supplied when the export job was created.

newExportJobProperties

Arguments

:: Text
$sel:jobId:ExportJobProperties'

-> JobStatus
$sel:jobStatus:ExportJobProperties'

-> UTCTime
$sel:submitTime:ExportJobProperties'

-> Text
$sel:datastoreId:ExportJobProperties'

-> OutputDataConfig
$sel:outputDataConfig:ExportJobProperties'

-> ExportJobProperties

Create a value of ExportJobProperties with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:jobName:ExportJobProperties', exportJobProperties_jobName - The user generated name for an export job.

$sel:endTime:ExportJobProperties', exportJobProperties_endTime - The time an export job completed.

$sel:dataAccessRoleArn:ExportJobProperties', exportJobProperties_dataAccessRoleArn - The Amazon Resource Name used during the initiation of the job.

$sel:message:ExportJobProperties', exportJobProperties_message - An explanation of any errors that may have occurred during the export job.

$sel:jobId:ExportJobProperties', exportJobProperties_jobId - The AWS generated ID for an export job.

$sel:jobStatus:ExportJobProperties', exportJobProperties_jobStatus - The status of a FHIR export job. Possible statuses are SUBMITTED, IN_PROGRESS, COMPLETED, or FAILED.

$sel:submitTime:ExportJobProperties', exportJobProperties_submitTime - The time an export job was initiated.

$sel:datastoreId:ExportJobProperties', exportJobProperties_datastoreId - The AWS generated ID for the Data Store from which files are being exported for an export job.

$sel:outputDataConfig:ExportJobProperties', exportJobProperties_outputDataConfig - The output data configuration that was supplied when the export job was created.

exportJobProperties_jobName :: Lens' ExportJobProperties (Maybe Text)

The user generated name for an export job.

exportJobProperties_endTime :: Lens' ExportJobProperties (Maybe UTCTime)

The time an export job completed.

exportJobProperties_dataAccessRoleArn :: Lens' ExportJobProperties (Maybe Text)

The Amazon Resource Name used during the initiation of the job.

exportJobProperties_message :: Lens' ExportJobProperties (Maybe Text)

An explanation of any errors that may have occurred during the export job.

exportJobProperties_jobId :: Lens' ExportJobProperties Text

The AWS generated ID for an export job.

exportJobProperties_jobStatus :: Lens' ExportJobProperties JobStatus

The status of a FHIR export job. Possible statuses are SUBMITTED, IN_PROGRESS, COMPLETED, or FAILED.

exportJobProperties_submitTime :: Lens' ExportJobProperties UTCTime

The time an export job was initiated.

exportJobProperties_datastoreId :: Lens' ExportJobProperties Text

The AWS generated ID for the Data Store from which files are being exported for an export job.

exportJobProperties_outputDataConfig :: Lens' ExportJobProperties OutputDataConfig

The output data configuration that was supplied when the export job was created.

ImportJobProperties

data ImportJobProperties

Displays the properties of the import job, including the ID, Arn, Name, and the status of the Data Store.

See: newImportJobProperties smart constructor.

Constructors

ImportJobProperties'

Fields

jobOutputDataConfig :: Maybe OutputDataConfig

jobName :: Maybe Text

The user-generated name for an Import job.

endTime :: Maybe POSIX

The time that the Import job was completed.

dataAccessRoleArn :: Maybe Text

The Amazon Resource Name (ARN) that gives Amazon HealthLake access to your input data.

message :: Maybe Text

An explanation of any errors that may have occurred during the FHIR import job.

jobId :: Text

The AWS-generated id number for the Import job.

jobStatus :: JobStatus

The job status for an Import job. Possible statuses are SUBMITTED, IN_PROGRESS, COMPLETED, FAILED.

submitTime :: POSIX

The time that the Import job was submitted for processing.

datastoreId :: Text

The datastore id used when the Import job was created.

inputDataConfig :: InputDataConfig

The input data configuration that was supplied when the Import job was created.

newImportJobProperties

Arguments

:: Text
$sel:jobId:ImportJobProperties'

-> JobStatus
$sel:jobStatus:ImportJobProperties'

-> UTCTime
$sel:submitTime:ImportJobProperties'

-> Text
$sel:datastoreId:ImportJobProperties'

-> InputDataConfig
$sel:inputDataConfig:ImportJobProperties'

-> ImportJobProperties

Create a value of ImportJobProperties with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:jobOutputDataConfig:ImportJobProperties', importJobProperties_jobOutputDataConfig - Undocumented member.

$sel:jobName:ImportJobProperties', importJobProperties_jobName - The user-generated name for an Import job.

$sel:endTime:ImportJobProperties', importJobProperties_endTime - The time that the Import job was completed.

$sel:dataAccessRoleArn:ImportJobProperties', importJobProperties_dataAccessRoleArn - The Amazon Resource Name (ARN) that gives Amazon HealthLake access to your input data.

$sel:message:ImportJobProperties', importJobProperties_message - An explanation of any errors that may have occurred during the FHIR import job.

$sel:jobId:ImportJobProperties', importJobProperties_jobId - The AWS-generated id number for the Import job.

$sel:jobStatus:ImportJobProperties', importJobProperties_jobStatus - The job status for an Import job. Possible statuses are SUBMITTED, IN_PROGRESS, COMPLETED, FAILED.

$sel:submitTime:ImportJobProperties', importJobProperties_submitTime - The time that the Import job was submitted for processing.

$sel:datastoreId:ImportJobProperties', importJobProperties_datastoreId - The datastore id used when the Import job was created.

$sel:inputDataConfig:ImportJobProperties', importJobProperties_inputDataConfig - The input data configuration that was supplied when the Import job was created.

importJobProperties_jobOutputDataConfig :: Lens' ImportJobProperties (Maybe OutputDataConfig)

Undocumented member.

importJobProperties_jobName :: Lens' ImportJobProperties (Maybe Text)

The user-generated name for an Import job.

importJobProperties_endTime :: Lens' ImportJobProperties (Maybe UTCTime)

The time that the Import job was completed.

importJobProperties_dataAccessRoleArn :: Lens' ImportJobProperties (Maybe Text)

The Amazon Resource Name (ARN) that gives Amazon HealthLake access to your input data.

importJobProperties_message :: Lens' ImportJobProperties (Maybe Text)

An explanation of any errors that may have occurred during the FHIR import job.

importJobProperties_jobId :: Lens' ImportJobProperties Text

The AWS-generated id number for the Import job.

importJobProperties_jobStatus :: Lens' ImportJobProperties JobStatus

The job status for an Import job. Possible statuses are SUBMITTED, IN_PROGRESS, COMPLETED, FAILED.

importJobProperties_submitTime :: Lens' ImportJobProperties UTCTime

The time that the Import job was submitted for processing.

importJobProperties_datastoreId :: Lens' ImportJobProperties Text

The datastore id used when the Import job was created.

importJobProperties_inputDataConfig :: Lens' ImportJobProperties InputDataConfig

The input data configuration that was supplied when the Import job was created.

InputDataConfig

data InputDataConfig

The input properties for an import job.

See: newInputDataConfig smart constructor.

Constructors

InputDataConfig'

Fields

s3Uri :: Maybe Text

The S3Uri is the user specified S3 location of the FHIR data to be imported into Amazon HealthLake.

newInputDataConfig :: InputDataConfig

Create a value of InputDataConfig with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:s3Uri:InputDataConfig', inputDataConfig_s3Uri - The S3Uri is the user specified S3 location of the FHIR data to be imported into Amazon HealthLake.

inputDataConfig_s3Uri :: Lens' InputDataConfig (Maybe Text)

The S3Uri is the user specified S3 location of the FHIR data to be imported into Amazon HealthLake.

KmsEncryptionConfig

data KmsEncryptionConfig

The customer-managed-key(CMK) used when creating a Data Store. If a customer owned key is not specified, an AWS owned key will be used for encryption.

See: newKmsEncryptionConfig smart constructor.

Constructors

KmsEncryptionConfig'

Fields

kmsKeyId :: Maybe Text

The KMS encryption key id/alias used to encrypt the Data Store contents at rest.

cmkType :: CmkType

The type of customer-managed-key(CMK) used for encyrption. The two types of supported CMKs are customer owned CMKs and AWS owned CMKs.

newKmsEncryptionConfig

Arguments

:: CmkType
$sel:cmkType:KmsEncryptionConfig'

-> KmsEncryptionConfig

Create a value of KmsEncryptionConfig with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:kmsKeyId:KmsEncryptionConfig', kmsEncryptionConfig_kmsKeyId - The KMS encryption key id/alias used to encrypt the Data Store contents at rest.

$sel:cmkType:KmsEncryptionConfig', kmsEncryptionConfig_cmkType - The type of customer-managed-key(CMK) used for encyrption. The two types of supported CMKs are customer owned CMKs and AWS owned CMKs.

kmsEncryptionConfig_kmsKeyId :: Lens' KmsEncryptionConfig (Maybe Text)

The KMS encryption key id/alias used to encrypt the Data Store contents at rest.

kmsEncryptionConfig_cmkType :: Lens' KmsEncryptionConfig CmkType

The type of customer-managed-key(CMK) used for encyrption. The two types of supported CMKs are customer owned CMKs and AWS owned CMKs.

OutputDataConfig

data OutputDataConfig

The output data configuration that was supplied when the export job was created.

See: newOutputDataConfig smart constructor.

Constructors

OutputDataConfig'

Fields

s3Configuration :: Maybe S3Configuration

The output data configuration that was supplied when the export job was created.

newOutputDataConfig :: OutputDataConfig

Create a value of OutputDataConfig with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:s3Configuration:OutputDataConfig', outputDataConfig_s3Configuration - The output data configuration that was supplied when the export job was created.

outputDataConfig_s3Configuration :: Lens' OutputDataConfig (Maybe S3Configuration)

The output data configuration that was supplied when the export job was created.

PreloadDataConfig

data PreloadDataConfig

The input properties for the preloaded Data Store. Only data preloaded from Synthea is supported.

See: newPreloadDataConfig smart constructor.

Constructors

PreloadDataConfig'

Fields

preloadDataType :: PreloadDataType

The type of preloaded data. Only Synthea preloaded data is supported.

newPreloadDataConfig

Arguments

:: PreloadDataType
$sel:preloadDataType:PreloadDataConfig'

-> PreloadDataConfig

Create a value of PreloadDataConfig with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:preloadDataType:PreloadDataConfig', preloadDataConfig_preloadDataType - The type of preloaded data. Only Synthea preloaded data is supported.

preloadDataConfig_preloadDataType :: Lens' PreloadDataConfig PreloadDataType

The type of preloaded data. Only Synthea preloaded data is supported.

S3Configuration

data S3Configuration

The configuration of the S3 bucket for either an import or export job. This includes assigning permissions for access.

See: newS3Configuration smart constructor.

Constructors

S3Configuration'

Fields

s3Uri :: Text

The S3Uri is the user specified S3 location of the FHIR data to be imported into Amazon HealthLake.

kmsKeyId :: Text

The KMS key ID used to access the S3 bucket.

newS3Configuration

Arguments

:: Text
$sel:s3Uri:S3Configuration'

-> Text
$sel:kmsKeyId:S3Configuration'

-> S3Configuration

Create a value of S3Configuration with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:s3Uri:S3Configuration', s3Configuration_s3Uri - The S3Uri is the user specified S3 location of the FHIR data to be imported into Amazon HealthLake.

$sel:kmsKeyId:S3Configuration', s3Configuration_kmsKeyId - The KMS key ID used to access the S3 bucket.

s3Configuration_s3Uri :: Lens' S3Configuration Text

The S3Uri is the user specified S3 location of the FHIR data to be imported into Amazon HealthLake.

s3Configuration_kmsKeyId :: Lens' S3Configuration Text

The KMS key ID used to access the S3 bucket.

SseConfiguration

data SseConfiguration

The server-side encryption key configuration for a customer provided encryption key.

See: newSseConfiguration smart constructor.

Constructors

SseConfiguration'

Fields

kmsEncryptionConfig :: KmsEncryptionConfig

The KMS encryption configuration used to provide details for data encryption.

newSseConfiguration

Arguments

:: KmsEncryptionConfig
$sel:kmsEncryptionConfig:SseConfiguration'

-> SseConfiguration

Create a value of SseConfiguration with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:kmsEncryptionConfig:SseConfiguration', sseConfiguration_kmsEncryptionConfig - The KMS encryption configuration used to provide details for data encryption.

sseConfiguration_kmsEncryptionConfig :: Lens' SseConfiguration KmsEncryptionConfig

The KMS encryption configuration used to provide details for data encryption.

Tag

data Tag

A tag is a label consisting of a user-defined key and value. The form for tags is {"Key", "Value"}

See: newTag smart constructor.

Constructors

Tag'

Fields

key :: Text

The key portion of a tag. Tag keys are case sensitive.

value :: Text

The value portion of tag. Tag values are case sensitive.

newTag

Arguments

:: Text
$sel:key:Tag'

-> Text
$sel:value:Tag'

-> Tag

Create a value of Tag with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:key:Tag', tag_key - The key portion of a tag. Tag keys are case sensitive.

$sel:value:Tag', tag_value - The value portion of tag. Tag values are case sensitive.

tag_key :: Lens' Tag Text

The key portion of a tag. Tag keys are case sensitive.

tag_value :: Lens' Tag Text

The value portion of tag. Tag values are case sensitive.