Amazonka.HealthLake.Types.DatastoreProperties

Description

Documentation

data DatastoreProperties

Displays the properties of the Data Store, including the ID, Arn, name, and the status of the Data Store.

See: newDatastoreProperties smart constructor.

Constructors

DatastoreProperties'

Fields

sseConfiguration :: Maybe SseConfiguration

The server-side encryption key configuration for a customer provided encryption key (CMK).

createdAt :: Maybe POSIX

The time that a Data Store was created.

datastoreName :: Maybe Text

The user-generated name for the Data Store.

preloadDataConfig :: Maybe PreloadDataConfig

The preloaded data configuration for the Data Store. Only data preloaded from Synthea is supported.

datastoreId :: Text

The AWS-generated ID number for the Data Store.

datastoreArn :: Text

The Amazon Resource Name used in the creation of the Data Store.

datastoreStatus :: DatastoreStatus

The status of the Data Store. Possible statuses are 'CREATING', 'ACTIVE', 'DELETING', or 'DELETED'.

datastoreTypeVersion :: FHIRVersion

The FHIR version. Only R4 version data is supported.

datastoreEndpoint :: Text

The AWS endpoint for the Data Store. Each Data Store will have it's own endpoint with Data Store ID in the endpoint URL.

newDatastoreProperties

Arguments

:: Text
$sel:datastoreId:DatastoreProperties'

-> Text
$sel:datastoreArn:DatastoreProperties'

-> DatastoreStatus
$sel:datastoreStatus:DatastoreProperties'

-> FHIRVersion
$sel:datastoreTypeVersion:DatastoreProperties'

-> Text
$sel:datastoreEndpoint:DatastoreProperties'

-> DatastoreProperties

Create a value of DatastoreProperties with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:sseConfiguration:DatastoreProperties', datastoreProperties_sseConfiguration - The server-side encryption key configuration for a customer provided encryption key (CMK).

$sel:createdAt:DatastoreProperties', datastoreProperties_createdAt - The time that a Data Store was created.

$sel:datastoreName:DatastoreProperties', datastoreProperties_datastoreName - The user-generated name for the Data Store.

$sel:preloadDataConfig:DatastoreProperties', datastoreProperties_preloadDataConfig - The preloaded data configuration for the Data Store. Only data preloaded from Synthea is supported.

$sel:datastoreId:DatastoreProperties', datastoreProperties_datastoreId - The AWS-generated ID number for the Data Store.

$sel:datastoreArn:DatastoreProperties', datastoreProperties_datastoreArn - The Amazon Resource Name used in the creation of the Data Store.

$sel:datastoreStatus:DatastoreProperties', datastoreProperties_datastoreStatus - The status of the Data Store. Possible statuses are 'CREATING', 'ACTIVE', 'DELETING', or 'DELETED'.

$sel:datastoreTypeVersion:DatastoreProperties', datastoreProperties_datastoreTypeVersion - The FHIR version. Only R4 version data is supported.

$sel:datastoreEndpoint:DatastoreProperties', datastoreProperties_datastoreEndpoint - The AWS endpoint for the Data Store. Each Data Store will have it's own endpoint with Data Store ID in the endpoint URL.

datastoreProperties_sseConfiguration :: Lens' DatastoreProperties (Maybe SseConfiguration)

The server-side encryption key configuration for a customer provided encryption key (CMK).

datastoreProperties_createdAt :: Lens' DatastoreProperties (Maybe UTCTime)

The time that a Data Store was created.

datastoreProperties_datastoreName :: Lens' DatastoreProperties (Maybe Text)

The user-generated name for the Data Store.

datastoreProperties_preloadDataConfig :: Lens' DatastoreProperties (Maybe PreloadDataConfig)

The preloaded data configuration for the Data Store. Only data preloaded from Synthea is supported.

datastoreProperties_datastoreId :: Lens' DatastoreProperties Text

The AWS-generated ID number for the Data Store.

datastoreProperties_datastoreArn :: Lens' DatastoreProperties Text

The Amazon Resource Name used in the creation of the Data Store.

datastoreProperties_datastoreStatus :: Lens' DatastoreProperties DatastoreStatus

The status of the Data Store. Possible statuses are 'CREATING', 'ACTIVE', 'DELETING', or 'DELETED'.

datastoreProperties_datastoreTypeVersion :: Lens' DatastoreProperties FHIRVersion

The FHIR version. Only R4 version data is supported.

datastoreProperties_datastoreEndpoint :: Lens' DatastoreProperties Text

The AWS endpoint for the Data Store. Each Data Store will have it's own endpoint with Data Store ID in the endpoint URL.