Amazonka.HoneyCode.BatchUpsertTableRows

Description

The BatchUpsertTableRows API allows you to upsert one or more rows in a table. The upsert operation takes a filter expression as input and evaluates it to find matching rows on the destination table. If matching rows are found, it will update the cells in the matching rows to new values specified in the request. If no matching rows are found, a new row is added at the end of the table and the cells in that row are set to the new values specified in the request.

You can specify the values to set in some or all of the columns in the table for the matching or newly appended rows. If a column is not explicitly specified for a particular row, then that column will not be updated for that row. To clear out the data in a specific cell, you need to set the value as an empty string ("").

Creating a Request

data BatchUpsertTableRows

See: newBatchUpsertTableRows smart constructor.

Constructors

BatchUpsertTableRows'

Fields

clientRequestToken :: Maybe Text

The request token for performing the update action. Request tokens help to identify duplicate requests. If a call times out or fails due to a transient error like a failed network connection, you can retry the call with the same request token. The service ensures that if the first call using that request token is successfully performed, the second call will not perform the action again.

Note that request tokens are valid only for a few minutes. You cannot use request tokens to dedupe requests spanning hours or days.

workbookId :: Text

The ID of the workbook where the rows are being upserted.

If a workbook with the specified id could not be found, this API throws ResourceNotFoundException.

tableId :: Text

The ID of the table where the rows are being upserted.

If a table with the specified id could not be found, this API throws ResourceNotFoundException.

rowsToUpsert :: [UpsertRowData]

The list of rows to upsert in the table. Each item in this list needs to have a batch item id to uniquely identify the element in the request, a filter expression to find the rows to update for that element and the cell values to set for each column in the upserted rows. You need to specify at least one item in this list.

Note that if one of the filter formulas in the request fails to evaluate because of an error or one of the column ids in any of the rows does not exist in the table, then the request fails and no updates are made to the table.

newBatchUpsertTableRows

Arguments

:: Text
$sel:workbookId:BatchUpsertTableRows'

-> Text
$sel:tableId:BatchUpsertTableRows'

-> BatchUpsertTableRows

Create a value of BatchUpsertTableRows with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:clientRequestToken:BatchUpsertTableRows', batchUpsertTableRows_clientRequestToken - The request token for performing the update action. Request tokens help to identify duplicate requests. If a call times out or fails due to a transient error like a failed network connection, you can retry the call with the same request token. The service ensures that if the first call using that request token is successfully performed, the second call will not perform the action again.

Note that request tokens are valid only for a few minutes. You cannot use request tokens to dedupe requests spanning hours or days.

$sel:workbookId:BatchUpsertTableRows', batchUpsertTableRows_workbookId - The ID of the workbook where the rows are being upserted.

If a workbook with the specified id could not be found, this API throws ResourceNotFoundException.

$sel:tableId:BatchUpsertTableRows', batchUpsertTableRows_tableId - The ID of the table where the rows are being upserted.

If a table with the specified id could not be found, this API throws ResourceNotFoundException.

$sel:rowsToUpsert:BatchUpsertTableRows', batchUpsertTableRows_rowsToUpsert - The list of rows to upsert in the table. Each item in this list needs to have a batch item id to uniquely identify the element in the request, a filter expression to find the rows to update for that element and the cell values to set for each column in the upserted rows. You need to specify at least one item in this list.

Note that if one of the filter formulas in the request fails to evaluate because of an error or one of the column ids in any of the rows does not exist in the table, then the request fails and no updates are made to the table.

Request Lenses

batchUpsertTableRows_clientRequestToken :: Lens' BatchUpsertTableRows (Maybe Text)

The request token for performing the update action. Request tokens help to identify duplicate requests. If a call times out or fails due to a transient error like a failed network connection, you can retry the call with the same request token. The service ensures that if the first call using that request token is successfully performed, the second call will not perform the action again.

Note that request tokens are valid only for a few minutes. You cannot use request tokens to dedupe requests spanning hours or days.

batchUpsertTableRows_workbookId :: Lens' BatchUpsertTableRows Text

The ID of the workbook where the rows are being upserted.

If a workbook with the specified id could not be found, this API throws ResourceNotFoundException.

batchUpsertTableRows_tableId :: Lens' BatchUpsertTableRows Text

The ID of the table where the rows are being upserted.

If a table with the specified id could not be found, this API throws ResourceNotFoundException.

batchUpsertTableRows_rowsToUpsert :: Lens' BatchUpsertTableRows [UpsertRowData]

The list of rows to upsert in the table. Each item in this list needs to have a batch item id to uniquely identify the element in the request, a filter expression to find the rows to update for that element and the cell values to set for each column in the upserted rows. You need to specify at least one item in this list.

Note that if one of the filter formulas in the request fails to evaluate because of an error or one of the column ids in any of the rows does not exist in the table, then the request fails and no updates are made to the table.

Destructuring the Response

data BatchUpsertTableRowsResponse

See: newBatchUpsertTableRowsResponse smart constructor.

Constructors

BatchUpsertTableRowsResponse'

Fields

failedBatchItems :: Maybe [FailedBatchItem]

The list of batch items in the request that could not be updated or appended in the table. Each element in this list contains one item from the request that could not be updated in the table along with the reason why that item could not be updated or appended.

httpStatus :: Int

The response's http status code.

rows :: HashMap Text UpsertRowsResult

A map with the batch item id as the key and the result of the upsert operation as the value. The result of the upsert operation specifies whether existing rows were updated or a new row was appended, along with the list of row ids that were affected.

workbookCursor :: Integer

The updated workbook cursor after updating or appending rows in the table.

newBatchUpsertTableRowsResponse

Arguments

:: Int
$sel:httpStatus:BatchUpsertTableRowsResponse'

-> Integer
$sel:workbookCursor:BatchUpsertTableRowsResponse'

-> BatchUpsertTableRowsResponse

Create a value of BatchUpsertTableRowsResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:failedBatchItems:BatchUpsertTableRowsResponse', batchUpsertTableRowsResponse_failedBatchItems - The list of batch items in the request that could not be updated or appended in the table. Each element in this list contains one item from the request that could not be updated in the table along with the reason why that item could not be updated or appended.

$sel:httpStatus:BatchUpsertTableRowsResponse', batchUpsertTableRowsResponse_httpStatus - The response's http status code.

$sel:rows:BatchUpsertTableRowsResponse', batchUpsertTableRowsResponse_rows - A map with the batch item id as the key and the result of the upsert operation as the value. The result of the upsert operation specifies whether existing rows were updated or a new row was appended, along with the list of row ids that were affected.

$sel:workbookCursor:BatchUpsertTableRowsResponse', batchUpsertTableRowsResponse_workbookCursor - The updated workbook cursor after updating or appending rows in the table.

Response Lenses

batchUpsertTableRowsResponse_failedBatchItems :: Lens' BatchUpsertTableRowsResponse (Maybe [FailedBatchItem])

The list of batch items in the request that could not be updated or appended in the table. Each element in this list contains one item from the request that could not be updated in the table along with the reason why that item could not be updated or appended.

batchUpsertTableRowsResponse_httpStatus :: Lens' BatchUpsertTableRowsResponse Int

The response's http status code.

batchUpsertTableRowsResponse_rows :: Lens' BatchUpsertTableRowsResponse (HashMap Text UpsertRowsResult)

A map with the batch item id as the key and the result of the upsert operation as the value. The result of the upsert operation specifies whether existing rows were updated or a new row was appended, along with the list of row ids that were affected.

batchUpsertTableRowsResponse_workbookCursor :: Lens' BatchUpsertTableRowsResponse Integer

The updated workbook cursor after updating or appending rows in the table.