Amazonka.HoneyCode.InvokeScreenAutomation

Description

The InvokeScreenAutomation API allows invoking an action defined in a screen in a Honeycode app. The API allows setting local variables, which can then be used in the automation being invoked. This allows automating the Honeycode app interactions to write, update or delete data in the workbook.

Creating a Request

data InvokeScreenAutomation

See: newInvokeScreenAutomation smart constructor.

Constructors

InvokeScreenAutomation'

Fields

rowId :: Maybe Text

The row ID for the automation if the automation is defined inside a block with source or list.

variables :: Maybe (Sensitive (HashMap Text (Sensitive VariableValue)))

Variables are specified as a map where the key is the name of the variable as defined on the screen. The value is an object which currently has only one property, rawValue, which holds the value of the variable to be passed to the screen. Any variables defined in a screen are required to be passed in the call.

clientRequestToken :: Maybe Text

The request token for performing the automation action. Request tokens help to identify duplicate requests. If a call times out or fails due to a transient error like a failed network connection, you can retry the call with the same request token. The service ensures that if the first call using that request token is successfully performed, the second call will return the response of the previous call rather than performing the action again.

Note that request tokens are valid only for a few minutes. You cannot use request tokens to dedupe requests spanning hours or days.

workbookId :: Text

The ID of the workbook that contains the screen automation.

appId :: Text

The ID of the app that contains the screen automation.

screenId :: Text

The ID of the screen that contains the screen automation.

screenAutomationId :: Text

The ID of the automation action to be performed.

newInvokeScreenAutomation

Arguments

:: Text
$sel:workbookId:InvokeScreenAutomation'

-> Text
$sel:appId:InvokeScreenAutomation'

-> Text
$sel:screenId:InvokeScreenAutomation'

-> Text
$sel:screenAutomationId:InvokeScreenAutomation'

-> InvokeScreenAutomation

Create a value of InvokeScreenAutomation with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:rowId:InvokeScreenAutomation', invokeScreenAutomation_rowId - The row ID for the automation if the automation is defined inside a block with source or list.

$sel:variables:InvokeScreenAutomation', invokeScreenAutomation_variables - Variables are specified as a map where the key is the name of the variable as defined on the screen. The value is an object which currently has only one property, rawValue, which holds the value of the variable to be passed to the screen. Any variables defined in a screen are required to be passed in the call.

$sel:clientRequestToken:InvokeScreenAutomation', invokeScreenAutomation_clientRequestToken - The request token for performing the automation action. Request tokens help to identify duplicate requests. If a call times out or fails due to a transient error like a failed network connection, you can retry the call with the same request token. The service ensures that if the first call using that request token is successfully performed, the second call will return the response of the previous call rather than performing the action again.

Note that request tokens are valid only for a few minutes. You cannot use request tokens to dedupe requests spanning hours or days.

$sel:workbookId:InvokeScreenAutomation', invokeScreenAutomation_workbookId - The ID of the workbook that contains the screen automation.

$sel:appId:InvokeScreenAutomation', invokeScreenAutomation_appId - The ID of the app that contains the screen automation.

$sel:screenId:InvokeScreenAutomation', invokeScreenAutomation_screenId - The ID of the screen that contains the screen automation.

$sel:screenAutomationId:InvokeScreenAutomation', invokeScreenAutomation_screenAutomationId - The ID of the automation action to be performed.

Request Lenses

invokeScreenAutomation_rowId :: Lens' InvokeScreenAutomation (Maybe Text)

The row ID for the automation if the automation is defined inside a block with source or list.

invokeScreenAutomation_variables :: Lens' InvokeScreenAutomation (Maybe (HashMap Text VariableValue))

Variables are specified as a map where the key is the name of the variable as defined on the screen. The value is an object which currently has only one property, rawValue, which holds the value of the variable to be passed to the screen. Any variables defined in a screen are required to be passed in the call.

invokeScreenAutomation_clientRequestToken :: Lens' InvokeScreenAutomation (Maybe Text)

The request token for performing the automation action. Request tokens help to identify duplicate requests. If a call times out or fails due to a transient error like a failed network connection, you can retry the call with the same request token. The service ensures that if the first call using that request token is successfully performed, the second call will return the response of the previous call rather than performing the action again.

Note that request tokens are valid only for a few minutes. You cannot use request tokens to dedupe requests spanning hours or days.

invokeScreenAutomation_workbookId :: Lens' InvokeScreenAutomation Text

The ID of the workbook that contains the screen automation.

invokeScreenAutomation_appId :: Lens' InvokeScreenAutomation Text

The ID of the app that contains the screen automation.

invokeScreenAutomation_screenId :: Lens' InvokeScreenAutomation Text

The ID of the screen that contains the screen automation.

invokeScreenAutomation_screenAutomationId :: Lens' InvokeScreenAutomation Text

The ID of the automation action to be performed.

Destructuring the Response

data InvokeScreenAutomationResponse

See: newInvokeScreenAutomationResponse smart constructor.

Constructors

InvokeScreenAutomationResponse'

Fields

httpStatus :: Int

The response's http status code.

workbookCursor :: Integer

The updated workbook cursor after performing the automation action.

newInvokeScreenAutomationResponse

Arguments

:: Int
$sel:httpStatus:InvokeScreenAutomationResponse'

-> Integer
$sel:workbookCursor:InvokeScreenAutomationResponse'

-> InvokeScreenAutomationResponse

Create a value of InvokeScreenAutomationResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:httpStatus:InvokeScreenAutomationResponse', invokeScreenAutomationResponse_httpStatus - The response's http status code.

$sel:workbookCursor:InvokeScreenAutomationResponse', invokeScreenAutomationResponse_workbookCursor - The updated workbook cursor after performing the automation action.

Response Lenses

invokeScreenAutomationResponse_httpStatus :: Lens' InvokeScreenAutomationResponse Int

The response's http status code.

invokeScreenAutomationResponse_workbookCursor :: Lens' InvokeScreenAutomationResponse Integer

The updated workbook cursor after performing the automation action.