Amazonka.HoneyCode.ListTableColumns

Description

The ListTableColumns API allows you to retrieve a list of all the columns in a table in a workbook.

This operation returns paginated results.

Creating a Request

data ListTableColumns

See: newListTableColumns smart constructor.

Constructors

ListTableColumns'

Fields

nextToken :: Maybe Text

This parameter is optional. If a nextToken is not specified, the API returns the first page of data.

Pagination tokens expire after 1 hour. If you use a token that was returned more than an hour back, the API will throw ValidationException.

workbookId :: Text

The ID of the workbook that contains the table whose columns are being retrieved.

If a workbook with the specified id could not be found, this API throws ResourceNotFoundException.

tableId :: Text

The ID of the table whose columns are being retrieved.

If a table with the specified id could not be found, this API throws ResourceNotFoundException.

newListTableColumns

Arguments

:: Text
$sel:workbookId:ListTableColumns'

-> Text
$sel:tableId:ListTableColumns'

-> ListTableColumns

Create a value of ListTableColumns with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:nextToken:ListTableColumns', listTableColumns_nextToken - This parameter is optional. If a nextToken is not specified, the API returns the first page of data.

Pagination tokens expire after 1 hour. If you use a token that was returned more than an hour back, the API will throw ValidationException.

$sel:workbookId:ListTableColumns', listTableColumns_workbookId - The ID of the workbook that contains the table whose columns are being retrieved.

If a workbook with the specified id could not be found, this API throws ResourceNotFoundException.

$sel:tableId:ListTableColumns', listTableColumns_tableId - The ID of the table whose columns are being retrieved.

If a table with the specified id could not be found, this API throws ResourceNotFoundException.

Request Lenses

listTableColumns_nextToken :: Lens' ListTableColumns (Maybe Text)

This parameter is optional. If a nextToken is not specified, the API returns the first page of data.

Pagination tokens expire after 1 hour. If you use a token that was returned more than an hour back, the API will throw ValidationException.

listTableColumns_workbookId :: Lens' ListTableColumns Text

The ID of the workbook that contains the table whose columns are being retrieved.

If a workbook with the specified id could not be found, this API throws ResourceNotFoundException.

listTableColumns_tableId :: Lens' ListTableColumns Text

The ID of the table whose columns are being retrieved.

If a table with the specified id could not be found, this API throws ResourceNotFoundException.

Destructuring the Response

data ListTableColumnsResponse

See: newListTableColumnsResponse smart constructor.

Constructors

ListTableColumnsResponse'

Fields

nextToken :: Maybe Text

Provides the pagination token to load the next page if there are more results matching the request. If a pagination token is not present in the response, it means that all data matching the request has been loaded.

workbookCursor :: Maybe Integer

Indicates the cursor of the workbook at which the data returned by this request is read. Workbook cursor keeps increasing with every update and the increments are not sequential.

httpStatus :: Int

The response's http status code.

tableColumns :: [TableColumn]

The list of columns in the table.

newListTableColumnsResponse

Arguments

:: Int
$sel:httpStatus:ListTableColumnsResponse'

-> ListTableColumnsResponse

Create a value of ListTableColumnsResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:nextToken:ListTableColumns', listTableColumnsResponse_nextToken - Provides the pagination token to load the next page if there are more results matching the request. If a pagination token is not present in the response, it means that all data matching the request has been loaded.

$sel:workbookCursor:ListTableColumnsResponse', listTableColumnsResponse_workbookCursor - Indicates the cursor of the workbook at which the data returned by this request is read. Workbook cursor keeps increasing with every update and the increments are not sequential.

$sel:httpStatus:ListTableColumnsResponse', listTableColumnsResponse_httpStatus - The response's http status code.

$sel:tableColumns:ListTableColumnsResponse', listTableColumnsResponse_tableColumns - The list of columns in the table.

Response Lenses

listTableColumnsResponse_nextToken :: Lens' ListTableColumnsResponse (Maybe Text)

Provides the pagination token to load the next page if there are more results matching the request. If a pagination token is not present in the response, it means that all data matching the request has been loaded.

listTableColumnsResponse_workbookCursor :: Lens' ListTableColumnsResponse (Maybe Integer)

Indicates the cursor of the workbook at which the data returned by this request is read. Workbook cursor keeps increasing with every update and the increments are not sequential.

listTableColumnsResponse_httpStatus :: Lens' ListTableColumnsResponse Int

The response's http status code.

listTableColumnsResponse_tableColumns :: Lens' ListTableColumnsResponse [TableColumn]

The list of columns in the table.