Amazonka.HoneyCode.ListTables

Description

The ListTables API allows you to retrieve a list of all the tables in a workbook.

This operation returns paginated results.

Creating a Request

data ListTables

See: newListTables smart constructor.

Constructors

ListTables'

Fields

nextToken :: Maybe Text

This parameter is optional. If a nextToken is not specified, the API returns the first page of data.

Pagination tokens expire after 1 hour. If you use a token that was returned more than an hour back, the API will throw ValidationException.

maxResults :: Maybe Natural

The maximum number of tables to return in each page of the results.

workbookId :: Text

The ID of the workbook whose tables are being retrieved.

If a workbook with the specified id could not be found, this API throws ResourceNotFoundException.

newListTables

Arguments

:: Text
$sel:workbookId:ListTables'

-> ListTables

Create a value of ListTables with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:nextToken:ListTables', listTables_nextToken - This parameter is optional. If a nextToken is not specified, the API returns the first page of data.

Pagination tokens expire after 1 hour. If you use a token that was returned more than an hour back, the API will throw ValidationException.

$sel:maxResults:ListTables', listTables_maxResults - The maximum number of tables to return in each page of the results.

$sel:workbookId:ListTables', listTables_workbookId - The ID of the workbook whose tables are being retrieved.

If a workbook with the specified id could not be found, this API throws ResourceNotFoundException.

Request Lenses

listTables_nextToken :: Lens' ListTables (Maybe Text)

This parameter is optional. If a nextToken is not specified, the API returns the first page of data.

Pagination tokens expire after 1 hour. If you use a token that was returned more than an hour back, the API will throw ValidationException.

listTables_maxResults :: Lens' ListTables (Maybe Natural)

The maximum number of tables to return in each page of the results.

listTables_workbookId :: Lens' ListTables Text

The ID of the workbook whose tables are being retrieved.

If a workbook with the specified id could not be found, this API throws ResourceNotFoundException.

Destructuring the Response

data ListTablesResponse

See: newListTablesResponse smart constructor.

Constructors

ListTablesResponse'

Fields

nextToken :: Maybe Text

Provides the pagination token to load the next page if there are more results matching the request. If a pagination token is not present in the response, it means that all data matching the request has been loaded.

workbookCursor :: Maybe Integer

Indicates the cursor of the workbook at which the data returned by this request is read. Workbook cursor keeps increasing with every update and the increments are not sequential.

httpStatus :: Int

The response's http status code.

tables :: [Table]

The list of tables in the workbook.

newListTablesResponse

Arguments

:: Int
$sel:httpStatus:ListTablesResponse'

-> ListTablesResponse

Create a value of ListTablesResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:nextToken:ListTables', listTablesResponse_nextToken - Provides the pagination token to load the next page if there are more results matching the request. If a pagination token is not present in the response, it means that all data matching the request has been loaded.

$sel:workbookCursor:ListTablesResponse', listTablesResponse_workbookCursor - Indicates the cursor of the workbook at which the data returned by this request is read. Workbook cursor keeps increasing with every update and the increments are not sequential.

$sel:httpStatus:ListTablesResponse', listTablesResponse_httpStatus - The response's http status code.

$sel:tables:ListTablesResponse', listTablesResponse_tables - The list of tables in the workbook.

Response Lenses

listTablesResponse_nextToken :: Lens' ListTablesResponse (Maybe Text)

Provides the pagination token to load the next page if there are more results matching the request. If a pagination token is not present in the response, it means that all data matching the request has been loaded.

listTablesResponse_workbookCursor :: Lens' ListTablesResponse (Maybe Integer)

Indicates the cursor of the workbook at which the data returned by this request is read. Workbook cursor keeps increasing with every update and the increments are not sequential.

listTablesResponse_httpStatus :: Lens' ListTablesResponse Int

The response's http status code.

listTablesResponse_tables :: Lens' ListTablesResponse [Table]

The list of tables in the workbook.