Amazonka.HoneyCode.QueryTableRows

Description

The QueryTableRows API allows you to use a filter formula to query for specific rows in a table.

This operation returns paginated results.

Creating a Request

data QueryTableRows

See: newQueryTableRows smart constructor.

Constructors

QueryTableRows'

Fields

nextToken :: Maybe Text

This parameter is optional. If a nextToken is not specified, the API returns the first page of data.

Pagination tokens expire after 1 hour. If you use a token that was returned more than an hour back, the API will throw ValidationException.

maxResults :: Maybe Natural

The maximum number of rows to return in each page of the results.

workbookId :: Text

The ID of the workbook whose table rows are being queried.

If a workbook with the specified id could not be found, this API throws ResourceNotFoundException.

tableId :: Text

The ID of the table whose rows are being queried.

If a table with the specified id could not be found, this API throws ResourceNotFoundException.

filterFormula :: Filter

An object that represents a filter formula along with the id of the context row under which the filter function needs to evaluate.

newQueryTableRows

Arguments

:: Text
$sel:workbookId:QueryTableRows'

-> Text
$sel:tableId:QueryTableRows'

-> Filter
$sel:filterFormula:QueryTableRows'

-> QueryTableRows

Create a value of QueryTableRows with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:nextToken:QueryTableRows', queryTableRows_nextToken - This parameter is optional. If a nextToken is not specified, the API returns the first page of data.

Pagination tokens expire after 1 hour. If you use a token that was returned more than an hour back, the API will throw ValidationException.

$sel:maxResults:QueryTableRows', queryTableRows_maxResults - The maximum number of rows to return in each page of the results.

$sel:workbookId:QueryTableRows', queryTableRows_workbookId - The ID of the workbook whose table rows are being queried.

If a workbook with the specified id could not be found, this API throws ResourceNotFoundException.

$sel:tableId:QueryTableRows', queryTableRows_tableId - The ID of the table whose rows are being queried.

If a table with the specified id could not be found, this API throws ResourceNotFoundException.

$sel:filterFormula:QueryTableRows', queryTableRows_filterFormula - An object that represents a filter formula along with the id of the context row under which the filter function needs to evaluate.

Request Lenses

queryTableRows_nextToken :: Lens' QueryTableRows (Maybe Text)

This parameter is optional. If a nextToken is not specified, the API returns the first page of data.

Pagination tokens expire after 1 hour. If you use a token that was returned more than an hour back, the API will throw ValidationException.

queryTableRows_maxResults :: Lens' QueryTableRows (Maybe Natural)

The maximum number of rows to return in each page of the results.

queryTableRows_workbookId :: Lens' QueryTableRows Text

The ID of the workbook whose table rows are being queried.

If a workbook with the specified id could not be found, this API throws ResourceNotFoundException.

queryTableRows_tableId :: Lens' QueryTableRows Text

The ID of the table whose rows are being queried.

If a table with the specified id could not be found, this API throws ResourceNotFoundException.

queryTableRows_filterFormula :: Lens' QueryTableRows Filter

An object that represents a filter formula along with the id of the context row under which the filter function needs to evaluate.

Destructuring the Response

data QueryTableRowsResponse

See: newQueryTableRowsResponse smart constructor.

Constructors

QueryTableRowsResponse'

Fields

nextToken :: Maybe Text

Provides the pagination token to load the next page if there are more results matching the request. If a pagination token is not present in the response, it means that all data matching the request has been loaded.

httpStatus :: Int

The response's http status code.

columnIds :: NonEmpty Text

The list of columns in the table whose row data is returned in the result.

rows :: [TableRow]

The list of rows in the table that match the query filter.

workbookCursor :: Integer

Indicates the cursor of the workbook at which the data returned by this request is read. Workbook cursor keeps increasing with every update and the increments are not sequential.

newQueryTableRowsResponse

Arguments

:: Int
$sel:httpStatus:QueryTableRowsResponse'

-> NonEmpty Text
$sel:columnIds:QueryTableRowsResponse'

-> Integer
$sel:workbookCursor:QueryTableRowsResponse'

-> QueryTableRowsResponse

Create a value of QueryTableRowsResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:nextToken:QueryTableRows', queryTableRowsResponse_nextToken - Provides the pagination token to load the next page if there are more results matching the request. If a pagination token is not present in the response, it means that all data matching the request has been loaded.

$sel:httpStatus:QueryTableRowsResponse', queryTableRowsResponse_httpStatus - The response's http status code.

$sel:columnIds:QueryTableRowsResponse', queryTableRowsResponse_columnIds - The list of columns in the table whose row data is returned in the result.

$sel:rows:QueryTableRowsResponse', queryTableRowsResponse_rows - The list of rows in the table that match the query filter.

$sel:workbookCursor:QueryTableRowsResponse', queryTableRowsResponse_workbookCursor - Indicates the cursor of the workbook at which the data returned by this request is read. Workbook cursor keeps increasing with every update and the increments are not sequential.

Response Lenses

queryTableRowsResponse_nextToken :: Lens' QueryTableRowsResponse (Maybe Text)

Provides the pagination token to load the next page if there are more results matching the request. If a pagination token is not present in the response, it means that all data matching the request has been loaded.

queryTableRowsResponse_httpStatus :: Lens' QueryTableRowsResponse Int

The response's http status code.

queryTableRowsResponse_columnIds :: Lens' QueryTableRowsResponse (NonEmpty Text)

The list of columns in the table whose row data is returned in the result.

queryTableRowsResponse_rows :: Lens' QueryTableRowsResponse [TableRow]

The list of rows in the table that match the query filter.

queryTableRowsResponse_workbookCursor :: Lens' QueryTableRowsResponse Integer

Indicates the cursor of the workbook at which the data returned by this request is read. Workbook cursor keeps increasing with every update and the increments are not sequential.