Amazonka.HoneyCode.Types

Description

Service Configuration

defaultService :: Service

API version 2020-03-01 of the Amazon Honeycode SDK configuration.

Errors

_ValidationException :: AsError a => Getting (First ServiceError) a ServiceError

Request is invalid. The message in the response contains details on why the request is invalid.

_AccessDeniedException :: AsError a => Getting (First ServiceError) a ServiceError

You do not have sufficient access to perform this action. Check that the workbook is owned by you and your IAM policy allows access to the resource in the request.

_AutomationExecutionTimeoutException :: AsError a => Getting (First ServiceError) a ServiceError

The automation execution timed out.

_ServiceQuotaExceededException :: AsError a => Getting (First ServiceError) a ServiceError

The request caused service quota to be breached.

_ThrottlingException :: AsError a => Getting (First ServiceError) a ServiceError

Tps(transactions per second) rate reached.

_AutomationExecutionException :: AsError a => Getting (First ServiceError) a ServiceError

The automation execution did not end successfully.

_RequestTimeoutException :: AsError a => Getting (First ServiceError) a ServiceError

The request timed out.

_InternalServerException :: AsError a => Getting (First ServiceError) a ServiceError

There were unexpected errors from the server.

_ServiceUnavailableException :: AsError a => Getting (First ServiceError) a ServiceError

Remote service is unreachable.

_ResourceNotFoundException :: AsError a => Getting (First ServiceError) a ServiceError

A Workbook, Table, App, Screen or Screen Automation was not found with the given ID.

Format

newtype Format

Constructors

Format'

Fields

fromFormat :: Text

Bundled Patterns

pattern Format_ACCOUNTING :: Format

pattern Format_AUTO :: Format

pattern Format_CONTACT :: Format

pattern Format_CURRENCY :: Format

pattern Format_DATE :: Format

pattern Format_DATE_TIME :: Format

pattern Format_NUMBER :: Format

pattern Format_PERCENTAGE :: Format

pattern Format_ROWLINK :: Format

pattern Format_TEXT :: Format

pattern Format_TIME :: Format

ImportDataCharacterEncoding

newtype ImportDataCharacterEncoding

Constructors

ImportDataCharacterEncoding'

Fields

fromImportDataCharacterEncoding :: Text

Bundled Patterns

pattern ImportDataCharacterEncoding_ISO_8859_1 :: ImportDataCharacterEncoding

pattern ImportDataCharacterEncoding_US_ASCII :: ImportDataCharacterEncoding

pattern ImportDataCharacterEncoding_UTF_16 :: ImportDataCharacterEncoding

pattern ImportDataCharacterEncoding_UTF_16BE :: ImportDataCharacterEncoding

pattern ImportDataCharacterEncoding_UTF_16LE :: ImportDataCharacterEncoding

pattern ImportDataCharacterEncoding_UTF_8 :: ImportDataCharacterEncoding

ImportSourceDataFormat

newtype ImportSourceDataFormat

Constructors

ImportSourceDataFormat'

Fields

fromImportSourceDataFormat :: Text

Bundled Patterns

pattern ImportSourceDataFormat_DELIMITED_TEXT :: ImportSourceDataFormat

TableDataImportJobStatus

newtype TableDataImportJobStatus

Constructors

TableDataImportJobStatus'

Fields

fromTableDataImportJobStatus :: Text

Bundled Patterns

pattern TableDataImportJobStatus_COMPLETED :: TableDataImportJobStatus

pattern TableDataImportJobStatus_FAILED :: TableDataImportJobStatus

pattern TableDataImportJobStatus_IN_PROGRESS :: TableDataImportJobStatus

pattern TableDataImportJobStatus_SUBMITTED :: TableDataImportJobStatus

UpsertAction

newtype UpsertAction

Constructors

UpsertAction'

Fields

fromUpsertAction :: Text

Bundled Patterns

pattern UpsertAction_APPENDED :: UpsertAction

pattern UpsertAction_UPDATED :: UpsertAction

Cell

data Cell

An object that represents a single cell in a table.

See: newCell smart constructor.

Constructors

Cell'

Fields

rawValue :: Maybe Text

The raw value of the data contained in the cell. The raw value depends on the format of the data in the cell. However the attribute in the API return value is always a string containing the raw value.

Cells with format DATE, DATE_TIME or TIME have the raw value as a floating point number where the whole number represents the number of days since 1/1/1900 and the fractional part represents the fraction of the day since midnight. For example, a cell with date 11/3/2020 has the raw value "44138". A cell with the time 9:00 AM has the raw value "0.375" and a cell with date/time value of 11/3/2020 9:00 AM has the raw value "44138.375". Notice that even though the raw value is a number in all three cases, it is still represented as a string.

Cells with format NUMBER, CURRENCY, PERCENTAGE and ACCOUNTING have the raw value of the data as the number representing the data being displayed. For example, the number 1.325 with two decimal places in the format will have it's raw value as "1.325" and formatted value as "1.33". A currency value for $10 will have the raw value as "10" and formatted value as "$10.00". A value representing 20% with two decimal places in the format will have its raw value as "0.2" and the formatted value as "20.00%". An accounting value of -$25 will have "-25" as the raw value and "$ (25.00)" as the formatted value.

Cells with format TEXT will have the raw text as the raw value. For example, a cell with text "John Smith" will have "John Smith" as both the raw value and the formatted value.

Cells with format CONTACT will have the name of the contact as a formatted value and the email address of the contact as the raw value. For example, a contact for John Smith will have "John Smith" as the formatted value and "john.smith@example.com" as the raw value.

Cells with format ROWLINK (aka picklist) will have the first column of the linked row as the formatted value and the row id of the linked row as the raw value. For example, a cell containing a picklist to a table that displays task status might have "Completed" as the formatted value and "row:dfcefaee-5b37-4355-8f28-40c3e4ff5dd4/ca432b2f-b8eb-431d-9fb5-cbe0342f9f03" as the raw value.

Cells with format AUTO or cells without any format that are auto-detected as one of the formats above will contain the raw and formatted values as mentioned above, based on the auto-detected formats. If there is no auto-detected format, the raw and formatted values will be the same as the data in the cell.

format :: Maybe Format

The format of the cell. If this field is empty, then the format is either not specified in the workbook or the format is set to AUTO.

formula :: Maybe (Sensitive Text)

The formula contained in the cell. This field is empty if a cell does not have a formula.

formattedValue :: Maybe Text

The formatted value of the cell. This is the value that you see displayed in the cell in the UI.

Note that the formatted value of a cell is always represented as a string irrespective of the data that is stored in the cell. For example, if a cell contains a date, the formatted value of the cell is the string representation of the formatted date being shown in the cell in the UI. See details in the rawValue field below for how cells of different formats will have different raw and formatted values.

newCell :: Cell

Create a value of Cell with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:rawValue:Cell', cell_rawValue - The raw value of the data contained in the cell. The raw value depends on the format of the data in the cell. However the attribute in the API return value is always a string containing the raw value.

Cells with format DATE, DATE_TIME or TIME have the raw value as a floating point number where the whole number represents the number of days since 1/1/1900 and the fractional part represents the fraction of the day since midnight. For example, a cell with date 11/3/2020 has the raw value "44138". A cell with the time 9:00 AM has the raw value "0.375" and a cell with date/time value of 11/3/2020 9:00 AM has the raw value "44138.375". Notice that even though the raw value is a number in all three cases, it is still represented as a string.

Cells with format NUMBER, CURRENCY, PERCENTAGE and ACCOUNTING have the raw value of the data as the number representing the data being displayed. For example, the number 1.325 with two decimal places in the format will have it's raw value as "1.325" and formatted value as "1.33". A currency value for $10 will have the raw value as "10" and formatted value as "$10.00". A value representing 20% with two decimal places in the format will have its raw value as "0.2" and the formatted value as "20.00%". An accounting value of -$25 will have "-25" as the raw value and "$ (25.00)" as the formatted value.

Cells with format TEXT will have the raw text as the raw value. For example, a cell with text "John Smith" will have "John Smith" as both the raw value and the formatted value.

Cells with format CONTACT will have the name of the contact as a formatted value and the email address of the contact as the raw value. For example, a contact for John Smith will have "John Smith" as the formatted value and "john.smith@example.com" as the raw value.

Cells with format ROWLINK (aka picklist) will have the first column of the linked row as the formatted value and the row id of the linked row as the raw value. For example, a cell containing a picklist to a table that displays task status might have "Completed" as the formatted value and "row:dfcefaee-5b37-4355-8f28-40c3e4ff5dd4/ca432b2f-b8eb-431d-9fb5-cbe0342f9f03" as the raw value.

Cells with format AUTO or cells without any format that are auto-detected as one of the formats above will contain the raw and formatted values as mentioned above, based on the auto-detected formats. If there is no auto-detected format, the raw and formatted values will be the same as the data in the cell.

$sel:format:Cell', cell_format - The format of the cell. If this field is empty, then the format is either not specified in the workbook or the format is set to AUTO.

$sel:formula:Cell', cell_formula - The formula contained in the cell. This field is empty if a cell does not have a formula.

$sel:formattedValue:Cell', cell_formattedValue - The formatted value of the cell. This is the value that you see displayed in the cell in the UI.

Note that the formatted value of a cell is always represented as a string irrespective of the data that is stored in the cell. For example, if a cell contains a date, the formatted value of the cell is the string representation of the formatted date being shown in the cell in the UI. See details in the rawValue field below for how cells of different formats will have different raw and formatted values.

cell_rawValue :: Lens' Cell (Maybe Text)

The raw value of the data contained in the cell. The raw value depends on the format of the data in the cell. However the attribute in the API return value is always a string containing the raw value.

Cells with format DATE, DATE_TIME or TIME have the raw value as a floating point number where the whole number represents the number of days since 1/1/1900 and the fractional part represents the fraction of the day since midnight. For example, a cell with date 11/3/2020 has the raw value "44138". A cell with the time 9:00 AM has the raw value "0.375" and a cell with date/time value of 11/3/2020 9:00 AM has the raw value "44138.375". Notice that even though the raw value is a number in all three cases, it is still represented as a string.

Cells with format NUMBER, CURRENCY, PERCENTAGE and ACCOUNTING have the raw value of the data as the number representing the data being displayed. For example, the number 1.325 with two decimal places in the format will have it's raw value as "1.325" and formatted value as "1.33". A currency value for $10 will have the raw value as "10" and formatted value as "$10.00". A value representing 20% with two decimal places in the format will have its raw value as "0.2" and the formatted value as "20.00%". An accounting value of -$25 will have "-25" as the raw value and "$ (25.00)" as the formatted value.

Cells with format TEXT will have the raw text as the raw value. For example, a cell with text "John Smith" will have "John Smith" as both the raw value and the formatted value.

Cells with format CONTACT will have the name of the contact as a formatted value and the email address of the contact as the raw value. For example, a contact for John Smith will have "John Smith" as the formatted value and "john.smith@example.com" as the raw value.

Cells with format ROWLINK (aka picklist) will have the first column of the linked row as the formatted value and the row id of the linked row as the raw value. For example, a cell containing a picklist to a table that displays task status might have "Completed" as the formatted value and "row:dfcefaee-5b37-4355-8f28-40c3e4ff5dd4/ca432b2f-b8eb-431d-9fb5-cbe0342f9f03" as the raw value.

Cells with format AUTO or cells without any format that are auto-detected as one of the formats above will contain the raw and formatted values as mentioned above, based on the auto-detected formats. If there is no auto-detected format, the raw and formatted values will be the same as the data in the cell.

cell_format :: Lens' Cell (Maybe Format)

The format of the cell. If this field is empty, then the format is either not specified in the workbook or the format is set to AUTO.

cell_formula :: Lens' Cell (Maybe Text)

The formula contained in the cell. This field is empty if a cell does not have a formula.

cell_formattedValue :: Lens' Cell (Maybe Text)

The formatted value of the cell. This is the value that you see displayed in the cell in the UI.

Note that the formatted value of a cell is always represented as a string irrespective of the data that is stored in the cell. For example, if a cell contains a date, the formatted value of the cell is the string representation of the formatted date being shown in the cell in the UI. See details in the rawValue field below for how cells of different formats will have different raw and formatted values.

CellInput

data CellInput

CellInput object contains the data needed to create or update cells in a table.

See: newCellInput smart constructor.

Constructors

CellInput'

Fields

fact :: Maybe (Sensitive Text)

Fact represents the data that is entered into a cell. This data can be free text or a formula. Formulas need to start with the equals (=) sign.

newCellInput :: CellInput

Create a value of CellInput with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:fact:CellInput', cellInput_fact - Fact represents the data that is entered into a cell. This data can be free text or a formula. Formulas need to start with the equals (=) sign.

cellInput_fact :: Lens' CellInput (Maybe Text)

Fact represents the data that is entered into a cell. This data can be free text or a formula. Formulas need to start with the equals (=) sign.

ColumnMetadata

data ColumnMetadata

Metadata for column in the table.

See: newColumnMetadata smart constructor.

Constructors

ColumnMetadata'

Fields

name :: Sensitive Text

The name of the column.

format :: Format

The format of the column.

newColumnMetadata

Arguments

:: Text
$sel:name:ColumnMetadata'

-> Format
$sel:format:ColumnMetadata'

-> ColumnMetadata

Create a value of ColumnMetadata with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:name:ColumnMetadata', columnMetadata_name - The name of the column.

$sel:format:ColumnMetadata', columnMetadata_format - The format of the column.

columnMetadata_name :: Lens' ColumnMetadata Text

The name of the column.

columnMetadata_format :: Lens' ColumnMetadata Format

The format of the column.

CreateRowData

data CreateRowData

Data needed to create a single row in a table as part of the BatchCreateTableRows request.

See: newCreateRowData smart constructor.

Constructors

CreateRowData'

Fields

batchItemId :: Text

An external identifier that represents the single row that is being created as part of the BatchCreateTableRows request. This can be any string that you can use to identify the row in the request. The BatchCreateTableRows API puts the batch item id in the results to allow you to link data in the request to data in the results.

cellsToCreate :: HashMap Text CellInput

A map representing the cells to create in the new row. The key is the column id of the cell and the value is the CellInput object that represents the data to set in that cell.

newCreateRowData

Arguments

:: Text
$sel:batchItemId:CreateRowData'

-> CreateRowData

Create a value of CreateRowData with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:batchItemId:CreateRowData', createRowData_batchItemId - An external identifier that represents the single row that is being created as part of the BatchCreateTableRows request. This can be any string that you can use to identify the row in the request. The BatchCreateTableRows API puts the batch item id in the results to allow you to link data in the request to data in the results.

$sel:cellsToCreate:CreateRowData', createRowData_cellsToCreate - A map representing the cells to create in the new row. The key is the column id of the cell and the value is the CellInput object that represents the data to set in that cell.

createRowData_batchItemId :: Lens' CreateRowData Text

An external identifier that represents the single row that is being created as part of the BatchCreateTableRows request. This can be any string that you can use to identify the row in the request. The BatchCreateTableRows API puts the batch item id in the results to allow you to link data in the request to data in the results.

createRowData_cellsToCreate :: Lens' CreateRowData (HashMap Text CellInput)

A map representing the cells to create in the new row. The key is the column id of the cell and the value is the CellInput object that represents the data to set in that cell.

DataItem

data DataItem

The data in a particular data cell defined on the screen.

See: newDataItem smart constructor.

Constructors

DataItem'

Fields

rawValue :: Maybe Text

The raw value of the data. e.g. jsmith@example.com

overrideFormat :: Maybe Format

The overrideFormat is optional and is specified only if a particular row of data has a different format for the data than the default format defined on the screen or the table.

formattedValue :: Maybe Text

The formatted value of the data. e.g. John Smith.

newDataItem :: DataItem

Create a value of DataItem with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:rawValue:DataItem', dataItem_rawValue - The raw value of the data. e.g. jsmith@example.com

$sel:overrideFormat:DataItem', dataItem_overrideFormat - The overrideFormat is optional and is specified only if a particular row of data has a different format for the data than the default format defined on the screen or the table.

$sel:formattedValue:DataItem', dataItem_formattedValue - The formatted value of the data. e.g. John Smith.

dataItem_rawValue :: Lens' DataItem (Maybe Text)

The raw value of the data. e.g. jsmith@example.com

dataItem_overrideFormat :: Lens' DataItem (Maybe Format)

The overrideFormat is optional and is specified only if a particular row of data has a different format for the data than the default format defined on the screen or the table.

dataItem_formattedValue :: Lens' DataItem (Maybe Text)

The formatted value of the data. e.g. John Smith.

DelimitedTextImportOptions

data DelimitedTextImportOptions

An object that contains the options relating to parsing delimited text as part of an import request.

See: newDelimitedTextImportOptions smart constructor.

Constructors

DelimitedTextImportOptions'

Fields

ignoreEmptyRows :: Maybe Bool

A parameter to indicate whether empty rows should be ignored or be included in the import.

hasHeaderRow :: Maybe Bool

Indicates whether the input file has a header row at the top containing the column names.

dataCharacterEncoding :: Maybe ImportDataCharacterEncoding

The encoding of the data in the input file.

delimiter :: Text

The delimiter to use for separating columns in a single row of the input.

newDelimitedTextImportOptions

Arguments

:: Text
$sel:delimiter:DelimitedTextImportOptions'

-> DelimitedTextImportOptions

Create a value of DelimitedTextImportOptions with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:ignoreEmptyRows:DelimitedTextImportOptions', delimitedTextImportOptions_ignoreEmptyRows - A parameter to indicate whether empty rows should be ignored or be included in the import.

$sel:hasHeaderRow:DelimitedTextImportOptions', delimitedTextImportOptions_hasHeaderRow - Indicates whether the input file has a header row at the top containing the column names.

$sel:dataCharacterEncoding:DelimitedTextImportOptions', delimitedTextImportOptions_dataCharacterEncoding - The encoding of the data in the input file.

$sel:delimiter:DelimitedTextImportOptions', delimitedTextImportOptions_delimiter - The delimiter to use for separating columns in a single row of the input.

delimitedTextImportOptions_ignoreEmptyRows :: Lens' DelimitedTextImportOptions (Maybe Bool)

A parameter to indicate whether empty rows should be ignored or be included in the import.

delimitedTextImportOptions_hasHeaderRow :: Lens' DelimitedTextImportOptions (Maybe Bool)

Indicates whether the input file has a header row at the top containing the column names.

delimitedTextImportOptions_dataCharacterEncoding :: Lens' DelimitedTextImportOptions (Maybe ImportDataCharacterEncoding)

The encoding of the data in the input file.

delimitedTextImportOptions_delimiter :: Lens' DelimitedTextImportOptions Text

The delimiter to use for separating columns in a single row of the input.

DestinationOptions

data DestinationOptions

An object that contains the options relating to the destination of the import request.

See: newDestinationOptions smart constructor.

Constructors

DestinationOptions'

Fields

columnMap :: Maybe (HashMap Text SourceDataColumnProperties)

A map of the column id to the import properties for each column.

newDestinationOptions :: DestinationOptions

Create a value of DestinationOptions with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:columnMap:DestinationOptions', destinationOptions_columnMap - A map of the column id to the import properties for each column.

destinationOptions_columnMap :: Lens' DestinationOptions (Maybe (HashMap Text SourceDataColumnProperties))

A map of the column id to the import properties for each column.

FailedBatchItem

data FailedBatchItem

A single item in a batch that failed to perform the intended action because of an error preventing it from succeeding.

See: newFailedBatchItem smart constructor.

Constructors

FailedBatchItem'

Fields

id :: Text

The id of the batch item that failed. This is the batch item id for the BatchCreateTableRows and BatchUpsertTableRows operations and the row id for the BatchUpdateTableRows and BatchDeleteTableRows operations.

errorMessage :: Text

The error message that indicates why the batch item failed.

newFailedBatchItem

Arguments

:: Text
$sel:id:FailedBatchItem'

-> Text
$sel:errorMessage:FailedBatchItem'

-> FailedBatchItem

Create a value of FailedBatchItem with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:id:FailedBatchItem', failedBatchItem_id - The id of the batch item that failed. This is the batch item id for the BatchCreateTableRows and BatchUpsertTableRows operations and the row id for the BatchUpdateTableRows and BatchDeleteTableRows operations.

$sel:errorMessage:FailedBatchItem', failedBatchItem_errorMessage - The error message that indicates why the batch item failed.

failedBatchItem_id :: Lens' FailedBatchItem Text

The id of the batch item that failed. This is the batch item id for the BatchCreateTableRows and BatchUpsertTableRows operations and the row id for the BatchUpdateTableRows and BatchDeleteTableRows operations.

failedBatchItem_errorMessage :: Lens' FailedBatchItem Text

The error message that indicates why the batch item failed.

Filter

data Filter

An object that represents a filter formula along with the id of the context row under which the filter function needs to evaluate.

See: newFilter smart constructor.

Constructors

Filter'

Fields

contextRowId :: Maybe Text

The optional contextRowId attribute can be used to specify the row id of the context row if the filter formula contains unqualified references to table columns and needs a context row to evaluate them successfully.

formula :: Sensitive Text

A formula representing a filter function that returns zero or more matching rows from a table. Valid formulas in this field return a list of rows from a table. The most common ways of writing a formula to return a list of rows are to use the FindRow() or Filter() functions. Any other formula that returns zero or more rows is also acceptable. For example, you can use a formula that points to a cell that contains a filter function.

newFilter

Arguments

:: Text
$sel:formula:Filter'

-> Filter

Create a value of Filter with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:contextRowId:Filter', filter_contextRowId - The optional contextRowId attribute can be used to specify the row id of the context row if the filter formula contains unqualified references to table columns and needs a context row to evaluate them successfully.

$sel:formula:Filter', filter_formula - A formula representing a filter function that returns zero or more matching rows from a table. Valid formulas in this field return a list of rows from a table. The most common ways of writing a formula to return a list of rows are to use the FindRow() or Filter() functions. Any other formula that returns zero or more rows is also acceptable. For example, you can use a formula that points to a cell that contains a filter function.

filter_contextRowId :: Lens' Filter (Maybe Text)

The optional contextRowId attribute can be used to specify the row id of the context row if the filter formula contains unqualified references to table columns and needs a context row to evaluate them successfully.

filter_formula :: Lens' Filter Text

A formula representing a filter function that returns zero or more matching rows from a table. Valid formulas in this field return a list of rows from a table. The most common ways of writing a formula to return a list of rows are to use the FindRow() or Filter() functions. Any other formula that returns zero or more rows is also acceptable. For example, you can use a formula that points to a cell that contains a filter function.

ImportDataSource

data ImportDataSource

An object that has details about the source of the data that was submitted for import.

See: newImportDataSource smart constructor.

Constructors

ImportDataSource'

Fields

dataSourceConfig :: ImportDataSourceConfig

The configuration parameters for the data source of the import

newImportDataSource

Arguments

:: ImportDataSourceConfig
$sel:dataSourceConfig:ImportDataSource'

-> ImportDataSource

Create a value of ImportDataSource with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:dataSourceConfig:ImportDataSource', importDataSource_dataSourceConfig - The configuration parameters for the data source of the import

importDataSource_dataSourceConfig :: Lens' ImportDataSource ImportDataSourceConfig

The configuration parameters for the data source of the import

ImportDataSourceConfig

data ImportDataSourceConfig

An object that contains the configuration parameters for the data source of an import request.

See: newImportDataSourceConfig smart constructor.

Constructors

ImportDataSourceConfig'

Fields

dataSourceUrl :: Maybe Text

The URL from which source data will be downloaded for the import request.

newImportDataSourceConfig :: ImportDataSourceConfig

Create a value of ImportDataSourceConfig with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:dataSourceUrl:ImportDataSourceConfig', importDataSourceConfig_dataSourceUrl - The URL from which source data will be downloaded for the import request.

importDataSourceConfig_dataSourceUrl :: Lens' ImportDataSourceConfig (Maybe Text)

The URL from which source data will be downloaded for the import request.

ImportJobSubmitter

data ImportJobSubmitter

An object that contains the attributes of the submitter of the import job.

See: newImportJobSubmitter smart constructor.

Constructors

ImportJobSubmitter'

Fields

email :: Maybe (Sensitive Text)

The email id of the submitter of the import job, if available.

userArn :: Maybe Text

The AWS user ARN of the submitter of the import job, if available.

newImportJobSubmitter :: ImportJobSubmitter

Create a value of ImportJobSubmitter with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:email:ImportJobSubmitter', importJobSubmitter_email - The email id of the submitter of the import job, if available.

$sel:userArn:ImportJobSubmitter', importJobSubmitter_userArn - The AWS user ARN of the submitter of the import job, if available.

importJobSubmitter_email :: Lens' ImportJobSubmitter (Maybe Text)

The email id of the submitter of the import job, if available.

importJobSubmitter_userArn :: Lens' ImportJobSubmitter (Maybe Text)

The AWS user ARN of the submitter of the import job, if available.

ImportOptions

data ImportOptions

An object that contains the options specified by the sumitter of the import request.

See: newImportOptions smart constructor.

Constructors

ImportOptions'

Fields

delimitedTextOptions :: Maybe DelimitedTextImportOptions

Options relating to parsing delimited text. Required if dataFormat is DELIMITED_TEXT.

destinationOptions :: Maybe DestinationOptions

Options relating to the destination of the import request.

newImportOptions :: ImportOptions

Create a value of ImportOptions with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:delimitedTextOptions:ImportOptions', importOptions_delimitedTextOptions - Options relating to parsing delimited text. Required if dataFormat is DELIMITED_TEXT.

$sel:destinationOptions:ImportOptions', importOptions_destinationOptions - Options relating to the destination of the import request.

importOptions_delimitedTextOptions :: Lens' ImportOptions (Maybe DelimitedTextImportOptions)

Options relating to parsing delimited text. Required if dataFormat is DELIMITED_TEXT.

importOptions_destinationOptions :: Lens' ImportOptions (Maybe DestinationOptions)

Options relating to the destination of the import request.

ResultRow

data ResultRow

A single row in the ResultSet.

See: newResultRow smart constructor.

Constructors

ResultRow'

Fields

rowId :: Maybe Text

The ID for a particular row.

dataItems :: [Sensitive DataItem]

List of all the data cells in a row.

newResultRow :: ResultRow

Create a value of ResultRow with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:rowId:ResultRow', resultRow_rowId - The ID for a particular row.

$sel:dataItems:ResultRow', resultRow_dataItems - List of all the data cells in a row.

resultRow_rowId :: Lens' ResultRow (Maybe Text)

The ID for a particular row.

resultRow_dataItems :: Lens' ResultRow [DataItem]

List of all the data cells in a row.

ResultSet

data ResultSet

ResultSet contains the results of the request for a single block or list defined on the screen.

See: newResultSet smart constructor.

Constructors

ResultSet'

Fields

headers :: [ColumnMetadata]

List of headers for all the data cells in the block. The header identifies the name and default format of the data cell. Data cells appear in the same order in all rows as defined in the header. The names and formats are not repeated in the rows. If a particular row does not have a value for a data cell, a blank value is used.

For example, a task list that displays the task name, due date and assigned person might have headers [ { "name": "Task Name"}, {"name": "Due Date", "format": "DATE"}, {"name": "Assigned", "format": "CONTACT"} ]. Every row in the result will have the task name as the first item, due date as the second item and assigned person as the third item. If a particular task does not have a due date, that row will still have a blank value in the second element and the assigned person will still be in the third element.

rows :: [ResultRow]

List of rows returned by the request. Each row has a row Id and a list of data cells in that row. The data cells will be present in the same order as they are defined in the header.

newResultSet :: ResultSet

Create a value of ResultSet with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:headers:ResultSet', resultSet_headers - List of headers for all the data cells in the block. The header identifies the name and default format of the data cell. Data cells appear in the same order in all rows as defined in the header. The names and formats are not repeated in the rows. If a particular row does not have a value for a data cell, a blank value is used.

For example, a task list that displays the task name, due date and assigned person might have headers [ { "name": "Task Name"}, {"name": "Due Date", "format": "DATE"}, {"name": "Assigned", "format": "CONTACT"} ]. Every row in the result will have the task name as the first item, due date as the second item and assigned person as the third item. If a particular task does not have a due date, that row will still have a blank value in the second element and the assigned person will still be in the third element.

$sel:rows:ResultSet', resultSet_rows - List of rows returned by the request. Each row has a row Id and a list of data cells in that row. The data cells will be present in the same order as they are defined in the header.

resultSet_headers :: Lens' ResultSet [ColumnMetadata]

List of headers for all the data cells in the block. The header identifies the name and default format of the data cell. Data cells appear in the same order in all rows as defined in the header. The names and formats are not repeated in the rows. If a particular row does not have a value for a data cell, a blank value is used.

For example, a task list that displays the task name, due date and assigned person might have headers [ { "name": "Task Name"}, {"name": "Due Date", "format": "DATE"}, {"name": "Assigned", "format": "CONTACT"} ]. Every row in the result will have the task name as the first item, due date as the second item and assigned person as the third item. If a particular task does not have a due date, that row will still have a blank value in the second element and the assigned person will still be in the third element.

resultSet_rows :: Lens' ResultSet [ResultRow]

List of rows returned by the request. Each row has a row Id and a list of data cells in that row. The data cells will be present in the same order as they are defined in the header.

SourceDataColumnProperties

data SourceDataColumnProperties

An object that contains the properties for importing data to a specific column in a table.

See: newSourceDataColumnProperties smart constructor.

Constructors

SourceDataColumnProperties'

Fields

columnIndex :: Maybe Natural

The index of the column in the input file.

newSourceDataColumnProperties :: SourceDataColumnProperties

Create a value of SourceDataColumnProperties with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:columnIndex:SourceDataColumnProperties', sourceDataColumnProperties_columnIndex - The index of the column in the input file.

sourceDataColumnProperties_columnIndex :: Lens' SourceDataColumnProperties (Maybe Natural)

The index of the column in the input file.

Table

data Table

An object representing the properties of a table in a workbook.

See: newTable smart constructor.

Constructors

Table'

Fields

tableId :: Maybe Text

The id of the table.

tableName :: Maybe Text

The name of the table.

newTable :: Table

Create a value of Table with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:tableId:Table', table_tableId - The id of the table.

$sel:tableName:Table', table_tableName - The name of the table.

table_tableId :: Lens' Table (Maybe Text)

The id of the table.

table_tableName :: Lens' Table (Maybe Text)

The name of the table.

TableColumn

data TableColumn

An object that contains attributes about a single column in a table

See: newTableColumn smart constructor.

Constructors

TableColumn'

Fields

format :: Maybe Format

The column level format that is applied in the table. An empty value in this field means that the column format is the default value 'AUTO'.

tableColumnName :: Maybe Text

The name of the column in the table.

tableColumnId :: Maybe Text

The id of the column in the table.

newTableColumn :: TableColumn

Create a value of TableColumn with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:format:TableColumn', tableColumn_format - The column level format that is applied in the table. An empty value in this field means that the column format is the default value 'AUTO'.

$sel:tableColumnName:TableColumn', tableColumn_tableColumnName - The name of the column in the table.

$sel:tableColumnId:TableColumn', tableColumn_tableColumnId - The id of the column in the table.

tableColumn_format :: Lens' TableColumn (Maybe Format)

The column level format that is applied in the table. An empty value in this field means that the column format is the default value 'AUTO'.

tableColumn_tableColumnName :: Lens' TableColumn (Maybe Text)

The name of the column in the table.

tableColumn_tableColumnId :: Lens' TableColumn (Maybe Text)

The id of the column in the table.

TableDataImportJobMetadata

data TableDataImportJobMetadata

The metadata associated with the table data import job that was submitted.

See: newTableDataImportJobMetadata smart constructor.

Constructors

TableDataImportJobMetadata'

Fields

submitter :: ImportJobSubmitter

Details about the submitter of the import request.

submitTime :: POSIX

The timestamp when the job was submitted for import.

importOptions :: ImportOptions

The options that was specified at the time of submitting the import request.

dataSource :: ImportDataSource

The source of the data that was submitted for import.

newTableDataImportJobMetadata

Arguments

:: ImportJobSubmitter
$sel:submitter:TableDataImportJobMetadata'

-> UTCTime
$sel:submitTime:TableDataImportJobMetadata'

-> ImportOptions
$sel:importOptions:TableDataImportJobMetadata'

-> ImportDataSource
$sel:dataSource:TableDataImportJobMetadata'

-> TableDataImportJobMetadata

Create a value of TableDataImportJobMetadata with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:submitter:TableDataImportJobMetadata', tableDataImportJobMetadata_submitter - Details about the submitter of the import request.

$sel:submitTime:TableDataImportJobMetadata', tableDataImportJobMetadata_submitTime - The timestamp when the job was submitted for import.

$sel:importOptions:TableDataImportJobMetadata', tableDataImportJobMetadata_importOptions - The options that was specified at the time of submitting the import request.

$sel:dataSource:TableDataImportJobMetadata', tableDataImportJobMetadata_dataSource - The source of the data that was submitted for import.

tableDataImportJobMetadata_submitter :: Lens' TableDataImportJobMetadata ImportJobSubmitter

Details about the submitter of the import request.

tableDataImportJobMetadata_submitTime :: Lens' TableDataImportJobMetadata UTCTime

The timestamp when the job was submitted for import.

tableDataImportJobMetadata_importOptions :: Lens' TableDataImportJobMetadata ImportOptions

The options that was specified at the time of submitting the import request.

tableDataImportJobMetadata_dataSource :: Lens' TableDataImportJobMetadata ImportDataSource

The source of the data that was submitted for import.

TableRow

data TableRow

An object that contains attributes about a single row in a table

See: newTableRow smart constructor.

Constructors

TableRow'

Fields

rowId :: Text

The id of the row in the table.

cells :: [Sensitive Cell]

A list of cells in the table row. The cells appear in the same order as the columns of the table.

newTableRow

Arguments

:: Text
$sel:rowId:TableRow'

-> TableRow

Create a value of TableRow with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:rowId:TableRow', tableRow_rowId - The id of the row in the table.

$sel:cells:TableRow', tableRow_cells - A list of cells in the table row. The cells appear in the same order as the columns of the table.

tableRow_rowId :: Lens' TableRow Text

The id of the row in the table.

tableRow_cells :: Lens' TableRow [Cell]

A list of cells in the table row. The cells appear in the same order as the columns of the table.

UpdateRowData

data UpdateRowData

Data needed to create a single row in a table as part of the BatchCreateTableRows request.

See: newUpdateRowData smart constructor.

Constructors

UpdateRowData'

Fields

rowId :: Text

The id of the row that needs to be updated.

cellsToUpdate :: HashMap Text CellInput

A map representing the cells to update in the given row. The key is the column id of the cell and the value is the CellInput object that represents the data to set in that cell.

newUpdateRowData

Arguments

:: Text
$sel:rowId:UpdateRowData'

-> UpdateRowData

Create a value of UpdateRowData with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:rowId:UpdateRowData', updateRowData_rowId - The id of the row that needs to be updated.

$sel:cellsToUpdate:UpdateRowData', updateRowData_cellsToUpdate - A map representing the cells to update in the given row. The key is the column id of the cell and the value is the CellInput object that represents the data to set in that cell.

updateRowData_rowId :: Lens' UpdateRowData Text

The id of the row that needs to be updated.

updateRowData_cellsToUpdate :: Lens' UpdateRowData (HashMap Text CellInput)

A map representing the cells to update in the given row. The key is the column id of the cell and the value is the CellInput object that represents the data to set in that cell.

UpsertRowData

data UpsertRowData

Data needed to upsert rows in a table as part of a single item in the BatchUpsertTableRows request.

See: newUpsertRowData smart constructor.

Constructors

UpsertRowData'

Fields

batchItemId :: Text

An external identifier that represents a single item in the request that is being upserted as part of the BatchUpsertTableRows request. This can be any string that you can use to identify the item in the request. The BatchUpsertTableRows API puts the batch item id in the results to allow you to link data in the request to data in the results.

filter' :: Filter

The filter formula to use to find existing matching rows to update. The formula needs to return zero or more rows. If the formula returns 0 rows, then a new row will be appended in the target table. If the formula returns one or more rows, then the returned rows will be updated.

Note that the filter formula needs to return rows from the target table for the upsert operation to succeed. If the filter formula has a syntax error or it doesn't evaluate to zero or more rows in the target table for any one item in the input list, then the entire BatchUpsertTableRows request fails and no updates are made to the table.

cellsToUpdate :: HashMap Text CellInput

A map representing the cells to update for the matching rows or an appended row. The key is the column id of the cell and the value is the CellInput object that represents the data to set in that cell.

newUpsertRowData

Arguments

:: Text
$sel:batchItemId:UpsertRowData'

-> Filter
$sel:filter':UpsertRowData'

-> UpsertRowData

Create a value of UpsertRowData with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:batchItemId:UpsertRowData', upsertRowData_batchItemId - An external identifier that represents a single item in the request that is being upserted as part of the BatchUpsertTableRows request. This can be any string that you can use to identify the item in the request. The BatchUpsertTableRows API puts the batch item id in the results to allow you to link data in the request to data in the results.

$sel:filter':UpsertRowData', upsertRowData_filter - The filter formula to use to find existing matching rows to update. The formula needs to return zero or more rows. If the formula returns 0 rows, then a new row will be appended in the target table. If the formula returns one or more rows, then the returned rows will be updated.

Note that the filter formula needs to return rows from the target table for the upsert operation to succeed. If the filter formula has a syntax error or it doesn't evaluate to zero or more rows in the target table for any one item in the input list, then the entire BatchUpsertTableRows request fails and no updates are made to the table.

$sel:cellsToUpdate:UpsertRowData', upsertRowData_cellsToUpdate - A map representing the cells to update for the matching rows or an appended row. The key is the column id of the cell and the value is the CellInput object that represents the data to set in that cell.

upsertRowData_batchItemId :: Lens' UpsertRowData Text

An external identifier that represents a single item in the request that is being upserted as part of the BatchUpsertTableRows request. This can be any string that you can use to identify the item in the request. The BatchUpsertTableRows API puts the batch item id in the results to allow you to link data in the request to data in the results.

upsertRowData_filter :: Lens' UpsertRowData Filter

The filter formula to use to find existing matching rows to update. The formula needs to return zero or more rows. If the formula returns 0 rows, then a new row will be appended in the target table. If the formula returns one or more rows, then the returned rows will be updated.

Note that the filter formula needs to return rows from the target table for the upsert operation to succeed. If the filter formula has a syntax error or it doesn't evaluate to zero or more rows in the target table for any one item in the input list, then the entire BatchUpsertTableRows request fails and no updates are made to the table.

upsertRowData_cellsToUpdate :: Lens' UpsertRowData (HashMap Text CellInput)

A map representing the cells to update for the matching rows or an appended row. The key is the column id of the cell and the value is the CellInput object that represents the data to set in that cell.

UpsertRowsResult

data UpsertRowsResult

An object that represents the result of a single upsert row request.

See: newUpsertRowsResult smart constructor.

Constructors

UpsertRowsResult'

Fields

rowIds :: NonEmpty Text

The list of row ids that were changed as part of an upsert row operation. If the upsert resulted in an update, this list could potentially contain multiple rows that matched the filter and hence got updated. If the upsert resulted in an append, this list would only have the single row that was appended.

upsertAction :: UpsertAction

The result of the upsert action.

newUpsertRowsResult

Arguments

:: NonEmpty Text
$sel:rowIds:UpsertRowsResult'

-> UpsertAction
$sel:upsertAction:UpsertRowsResult'

-> UpsertRowsResult

Create a value of UpsertRowsResult with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:rowIds:UpsertRowsResult', upsertRowsResult_rowIds - The list of row ids that were changed as part of an upsert row operation. If the upsert resulted in an update, this list could potentially contain multiple rows that matched the filter and hence got updated. If the upsert resulted in an append, this list would only have the single row that was appended.

$sel:upsertAction:UpsertRowsResult', upsertRowsResult_upsertAction - The result of the upsert action.

upsertRowsResult_rowIds :: Lens' UpsertRowsResult (NonEmpty Text)

The list of row ids that were changed as part of an upsert row operation. If the upsert resulted in an update, this list could potentially contain multiple rows that matched the filter and hence got updated. If the upsert resulted in an append, this list would only have the single row that was appended.

upsertRowsResult_upsertAction :: Lens' UpsertRowsResult UpsertAction

The result of the upsert action.

VariableValue

data VariableValue

The input variables to the app to be used by the InvokeScreenAutomation action request.

See: newVariableValue smart constructor.

Constructors

VariableValue'

Fields

rawValue :: Text

Raw value of the variable.

newVariableValue

Arguments

:: Text
$sel:rawValue:VariableValue'

-> VariableValue

Create a value of VariableValue with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:rawValue:VariableValue', variableValue_rawValue - Raw value of the variable.

variableValue_rawValue :: Lens' VariableValue Text

Raw value of the variable.