Amazonka.HoneyCode.Types.DelimitedTextImportOptions

Description

Documentation

data DelimitedTextImportOptions

An object that contains the options relating to parsing delimited text as part of an import request.

See: newDelimitedTextImportOptions smart constructor.

Constructors

DelimitedTextImportOptions'

Fields

ignoreEmptyRows :: Maybe Bool

A parameter to indicate whether empty rows should be ignored or be included in the import.

hasHeaderRow :: Maybe Bool

Indicates whether the input file has a header row at the top containing the column names.

dataCharacterEncoding :: Maybe ImportDataCharacterEncoding

The encoding of the data in the input file.

delimiter :: Text

The delimiter to use for separating columns in a single row of the input.

newDelimitedTextImportOptions

Arguments

:: Text
$sel:delimiter:DelimitedTextImportOptions'

-> DelimitedTextImportOptions

Create a value of DelimitedTextImportOptions with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:ignoreEmptyRows:DelimitedTextImportOptions', delimitedTextImportOptions_ignoreEmptyRows - A parameter to indicate whether empty rows should be ignored or be included in the import.

$sel:hasHeaderRow:DelimitedTextImportOptions', delimitedTextImportOptions_hasHeaderRow - Indicates whether the input file has a header row at the top containing the column names.

$sel:dataCharacterEncoding:DelimitedTextImportOptions', delimitedTextImportOptions_dataCharacterEncoding - The encoding of the data in the input file.

$sel:delimiter:DelimitedTextImportOptions', delimitedTextImportOptions_delimiter - The delimiter to use for separating columns in a single row of the input.

delimitedTextImportOptions_ignoreEmptyRows :: Lens' DelimitedTextImportOptions (Maybe Bool)

A parameter to indicate whether empty rows should be ignored or be included in the import.

delimitedTextImportOptions_hasHeaderRow :: Lens' DelimitedTextImportOptions (Maybe Bool)

Indicates whether the input file has a header row at the top containing the column names.

delimitedTextImportOptions_dataCharacterEncoding :: Lens' DelimitedTextImportOptions (Maybe ImportDataCharacterEncoding)

The encoding of the data in the input file.

delimitedTextImportOptions_delimiter :: Lens' DelimitedTextImportOptions Text

The delimiter to use for separating columns in a single row of the input.