Amazonka.HoneyCode.Types.TableDataImportJobMetadata

Description

Documentation

data TableDataImportJobMetadata

The metadata associated with the table data import job that was submitted.

See: newTableDataImportJobMetadata smart constructor.

Constructors

TableDataImportJobMetadata'

Fields

submitter :: ImportJobSubmitter

Details about the submitter of the import request.

submitTime :: POSIX

The timestamp when the job was submitted for import.

importOptions :: ImportOptions

The options that was specified at the time of submitting the import request.

dataSource :: ImportDataSource

The source of the data that was submitted for import.

newTableDataImportJobMetadata

Arguments

:: ImportJobSubmitter
$sel:submitter:TableDataImportJobMetadata'

-> UTCTime
$sel:submitTime:TableDataImportJobMetadata'

-> ImportOptions
$sel:importOptions:TableDataImportJobMetadata'

-> ImportDataSource
$sel:dataSource:TableDataImportJobMetadata'

-> TableDataImportJobMetadata

Create a value of TableDataImportJobMetadata with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:submitter:TableDataImportJobMetadata', tableDataImportJobMetadata_submitter - Details about the submitter of the import request.

$sel:submitTime:TableDataImportJobMetadata', tableDataImportJobMetadata_submitTime - The timestamp when the job was submitted for import.

$sel:importOptions:TableDataImportJobMetadata', tableDataImportJobMetadata_importOptions - The options that was specified at the time of submitting the import request.

$sel:dataSource:TableDataImportJobMetadata', tableDataImportJobMetadata_dataSource - The source of the data that was submitted for import.

tableDataImportJobMetadata_submitter :: Lens' TableDataImportJobMetadata ImportJobSubmitter

Details about the submitter of the import request.

tableDataImportJobMetadata_submitTime :: Lens' TableDataImportJobMetadata UTCTime

The timestamp when the job was submitted for import.

tableDataImportJobMetadata_importOptions :: Lens' TableDataImportJobMetadata ImportOptions

The options that was specified at the time of submitting the import request.

tableDataImportJobMetadata_dataSource :: Lens' TableDataImportJobMetadata ImportDataSource

The source of the data that was submitted for import.