Amazonka.HoneyCode.Types.TableRow

Description

Documentation

data TableRow

An object that contains attributes about a single row in a table

See: newTableRow smart constructor.

Constructors

TableRow'

Fields

rowId :: Text

The id of the row in the table.

cells :: [Sensitive Cell]

A list of cells in the table row. The cells appear in the same order as the columns of the table.

newTableRow

Arguments

:: Text
$sel:rowId:TableRow'

-> TableRow

Create a value of TableRow with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:rowId:TableRow', tableRow_rowId - The id of the row in the table.

$sel:cells:TableRow', tableRow_cells - A list of cells in the table row. The cells appear in the same order as the columns of the table.

tableRow_rowId :: Lens' TableRow Text

The id of the row in the table.

tableRow_cells :: Lens' TableRow [Cell]

A list of cells in the table row. The cells appear in the same order as the columns of the table.