Amazonka.HoneyCode.Types.UpdateRowData

Description

Documentation

data UpdateRowData

Data needed to create a single row in a table as part of the BatchCreateTableRows request.

See: newUpdateRowData smart constructor.

Constructors

UpdateRowData'

Fields

rowId :: Text

The id of the row that needs to be updated.

cellsToUpdate :: HashMap Text CellInput

A map representing the cells to update in the given row. The key is the column id of the cell and the value is the CellInput object that represents the data to set in that cell.

newUpdateRowData

Arguments

:: Text
$sel:rowId:UpdateRowData'

-> UpdateRowData

Create a value of UpdateRowData with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:rowId:UpdateRowData', updateRowData_rowId - The id of the row that needs to be updated.

$sel:cellsToUpdate:UpdateRowData', updateRowData_cellsToUpdate - A map representing the cells to update in the given row. The key is the column id of the cell and the value is the CellInput object that represents the data to set in that cell.

updateRowData_rowId :: Lens' UpdateRowData Text

The id of the row that needs to be updated.

updateRowData_cellsToUpdate :: Lens' UpdateRowData (HashMap Text CellInput)

A map representing the cells to update in the given row. The key is the column id of the cell and the value is the CellInput object that represents the data to set in that cell.