Amazonka.IAM.ChangePassword

Description

Changes the password of the IAM user who is calling this operation. This operation can be performed using the CLI, the Amazon Web Services API, or the My Security Credentials page in the Amazon Web Services Management Console. The Amazon Web Services account root user password is not affected by this operation.

Use UpdateLoginProfile to use the CLI, the Amazon Web Services API, or the Users page in the IAM console to change the password for any IAM user. For more information about modifying passwords, see Managing passwords in the IAM User Guide.

Creating a Request

data ChangePassword

See: newChangePassword smart constructor.

Constructors

ChangePassword'

Fields

oldPassword :: Sensitive Text

The IAM user's current password.

newPassword' :: Sensitive Text

The new password. The new password must conform to the Amazon Web Services account's password policy, if one exists.

The regex pattern that is used to validate this parameter is a string of characters. That string can include almost any printable ASCII character from the space (\u0020) through the end of the ASCII character range (\u00FF). You can also include the tab (\u0009), line feed (\u000A), and carriage return (\u000D) characters. Any of these characters are valid in a password. However, many tools, such as the Amazon Web Services Management Console, might restrict the ability to type certain characters because they have special meaning within that tool.

newChangePassword

Arguments

:: Text
$sel:oldPassword:ChangePassword'

-> Text
$sel:newPassword':ChangePassword'

-> ChangePassword

Create a value of ChangePassword with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:oldPassword:ChangePassword', changePassword_oldPassword - The IAM user's current password.

$sel:newPassword':ChangePassword', changePassword_newPassword - The new password. The new password must conform to the Amazon Web Services account's password policy, if one exists.

The regex pattern that is used to validate this parameter is a string of characters. That string can include almost any printable ASCII character from the space (\u0020) through the end of the ASCII character range (\u00FF). You can also include the tab (\u0009), line feed (\u000A), and carriage return (\u000D) characters. Any of these characters are valid in a password. However, many tools, such as the Amazon Web Services Management Console, might restrict the ability to type certain characters because they have special meaning within that tool.

Request Lenses

changePassword_oldPassword :: Lens' ChangePassword Text

The IAM user's current password.

changePassword_newPassword :: Lens' ChangePassword Text

The new password. The new password must conform to the Amazon Web Services account's password policy, if one exists.

The regex pattern that is used to validate this parameter is a string of characters. That string can include almost any printable ASCII character from the space (\u0020) through the end of the ASCII character range (\u00FF). You can also include the tab (\u0009), line feed (\u000A), and carriage return (\u000D) characters. Any of these characters are valid in a password. However, many tools, such as the Amazon Web Services Management Console, might restrict the ability to type certain characters because they have special meaning within that tool.

Destructuring the Response

data ChangePasswordResponse

See: newChangePasswordResponse smart constructor.

Constructors

ChangePasswordResponse'

newChangePasswordResponse :: ChangePasswordResponse

Create a value of ChangePasswordResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.