Amazonka.IAM.CreateGroup

Description

Creates a new group.

For information about the number of groups you can create, see IAM and STS quotas in the IAM User Guide.

Creating a Request

data CreateGroup

See: newCreateGroup smart constructor.

Constructors

CreateGroup'

Fields

path :: Maybe Text

The path to the group. For more information about paths, see IAM identifiers in the IAM User Guide.

This parameter is optional. If it is not included, it defaults to a slash (/).

This parameter allows (through its regex pattern) a string of characters consisting of either a forward slash (/) by itself or a string that must begin and end with forward slashes. In addition, it can contain any ASCII character from the ! (\u0021) through the DEL character (\u007F), including most punctuation characters, digits, and upper and lowercased letters.

groupName :: Text

The name of the group to create. Do not include the path in this value.

IAM user, group, role, and policy names must be unique within the account. Names are not distinguished by case. For example, you cannot create resources named both "MyResource" and "myresource".

newCreateGroup

Arguments

:: Text
$sel:groupName:CreateGroup'

-> CreateGroup

Create a value of CreateGroup with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:path:CreateGroup', createGroup_path - The path to the group. For more information about paths, see IAM identifiers in the IAM User Guide.

This parameter is optional. If it is not included, it defaults to a slash (/).

This parameter allows (through its regex pattern) a string of characters consisting of either a forward slash (/) by itself or a string that must begin and end with forward slashes. In addition, it can contain any ASCII character from the ! (\u0021) through the DEL character (\u007F), including most punctuation characters, digits, and upper and lowercased letters.

$sel:groupName:CreateGroup', createGroup_groupName - The name of the group to create. Do not include the path in this value.

IAM user, group, role, and policy names must be unique within the account. Names are not distinguished by case. For example, you cannot create resources named both "MyResource" and "myresource".

Request Lenses

createGroup_path :: Lens' CreateGroup (Maybe Text)

The path to the group. For more information about paths, see IAM identifiers in the IAM User Guide.

This parameter is optional. If it is not included, it defaults to a slash (/).

This parameter allows (through its regex pattern) a string of characters consisting of either a forward slash (/) by itself or a string that must begin and end with forward slashes. In addition, it can contain any ASCII character from the ! (\u0021) through the DEL character (\u007F), including most punctuation characters, digits, and upper and lowercased letters.

createGroup_groupName :: Lens' CreateGroup Text

The name of the group to create. Do not include the path in this value.

IAM user, group, role, and policy names must be unique within the account. Names are not distinguished by case. For example, you cannot create resources named both "MyResource" and "myresource".

Destructuring the Response

data CreateGroupResponse

Contains the response to a successful CreateGroup request.

See: newCreateGroupResponse smart constructor.

Constructors

CreateGroupResponse'

Fields

httpStatus :: Int

The response's http status code.

group' :: Group

A structure containing details about the new group.

newCreateGroupResponse

Arguments

:: Int
$sel:httpStatus:CreateGroupResponse'

-> Group
$sel:group':CreateGroupResponse'

-> CreateGroupResponse

Create a value of CreateGroupResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:httpStatus:CreateGroupResponse', createGroupResponse_httpStatus - The response's http status code.

$sel:group':CreateGroupResponse', createGroupResponse_group - A structure containing details about the new group.

Response Lenses

createGroupResponse_httpStatus :: Lens' CreateGroupResponse Int

The response's http status code.

createGroupResponse_group :: Lens' CreateGroupResponse Group

A structure containing details about the new group.