Amazonka.IAM.CreateInstanceProfile

Description

Creates a new instance profile. For information about instance profiles, see Using roles for applications on Amazon EC2 in the IAM User Guide, and Instance profiles in the Amazon EC2 User Guide.

For information about the number of instance profiles you can create, see IAM object quotas in the IAM User Guide.

Creating a Request

data CreateInstanceProfile

See: newCreateInstanceProfile smart constructor.

Constructors

CreateInstanceProfile'

Fields

path :: Maybe Text

The path to the instance profile. For more information about paths, see IAM Identifiers in the IAM User Guide.

This parameter is optional. If it is not included, it defaults to a slash (/).

This parameter allows (through its regex pattern) a string of characters consisting of either a forward slash (/) by itself or a string that must begin and end with forward slashes. In addition, it can contain any ASCII character from the ! (\u0021) through the DEL character (\u007F), including most punctuation characters, digits, and upper and lowercased letters.

tags :: Maybe [Tag]

A list of tags that you want to attach to the newly created IAM instance profile. Each tag consists of a key name and an associated value. For more information about tagging, see Tagging IAM resources in the IAM User Guide.

If any one of the tags is invalid or if you exceed the allowed maximum number of tags, then the entire request fails and the resource is not created.

instanceProfileName :: Text

The name of the instance profile to create.

This parameter allows (through its regex pattern) a string of characters consisting of upper and lowercase alphanumeric characters with no spaces. You can also include any of the following characters: _+=,.@-

newCreateInstanceProfile

Arguments

:: Text
$sel:instanceProfileName:CreateInstanceProfile'

-> CreateInstanceProfile

Create a value of CreateInstanceProfile with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:path:CreateInstanceProfile', createInstanceProfile_path - The path to the instance profile. For more information about paths, see IAM Identifiers in the IAM User Guide.

This parameter is optional. If it is not included, it defaults to a slash (/).

This parameter allows (through its regex pattern) a string of characters consisting of either a forward slash (/) by itself or a string that must begin and end with forward slashes. In addition, it can contain any ASCII character from the ! (\u0021) through the DEL character (\u007F), including most punctuation characters, digits, and upper and lowercased letters.

$sel:tags:CreateInstanceProfile', createInstanceProfile_tags - A list of tags that you want to attach to the newly created IAM instance profile. Each tag consists of a key name and an associated value. For more information about tagging, see Tagging IAM resources in the IAM User Guide.

If any one of the tags is invalid or if you exceed the allowed maximum number of tags, then the entire request fails and the resource is not created.

$sel:instanceProfileName:CreateInstanceProfile', createInstanceProfile_instanceProfileName - The name of the instance profile to create.

This parameter allows (through its regex pattern) a string of characters consisting of upper and lowercase alphanumeric characters with no spaces. You can also include any of the following characters: _+=,.@-

Request Lenses

createInstanceProfile_path :: Lens' CreateInstanceProfile (Maybe Text)

The path to the instance profile. For more information about paths, see IAM Identifiers in the IAM User Guide.

This parameter is optional. If it is not included, it defaults to a slash (/).

This parameter allows (through its regex pattern) a string of characters consisting of either a forward slash (/) by itself or a string that must begin and end with forward slashes. In addition, it can contain any ASCII character from the ! (\u0021) through the DEL character (\u007F), including most punctuation characters, digits, and upper and lowercased letters.

createInstanceProfile_tags :: Lens' CreateInstanceProfile (Maybe [Tag])

A list of tags that you want to attach to the newly created IAM instance profile. Each tag consists of a key name and an associated value. For more information about tagging, see Tagging IAM resources in the IAM User Guide.

If any one of the tags is invalid or if you exceed the allowed maximum number of tags, then the entire request fails and the resource is not created.

createInstanceProfile_instanceProfileName :: Lens' CreateInstanceProfile Text

The name of the instance profile to create.

This parameter allows (through its regex pattern) a string of characters consisting of upper and lowercase alphanumeric characters with no spaces. You can also include any of the following characters: _+=,.@-

Destructuring the Response

data CreateInstanceProfileResponse

Contains the response to a successful CreateInstanceProfile request.

See: newCreateInstanceProfileResponse smart constructor.

Constructors

CreateInstanceProfileResponse'

Fields

httpStatus :: Int

The response's http status code.

instanceProfile :: InstanceProfile

A structure containing details about the new instance profile.

newCreateInstanceProfileResponse

Arguments

:: Int
$sel:httpStatus:CreateInstanceProfileResponse'

-> InstanceProfile
$sel:instanceProfile:CreateInstanceProfileResponse'

-> CreateInstanceProfileResponse

Create a value of CreateInstanceProfileResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:httpStatus:CreateInstanceProfileResponse', createInstanceProfileResponse_httpStatus - The response's http status code.

$sel:instanceProfile:CreateInstanceProfileResponse', createInstanceProfileResponse_instanceProfile - A structure containing details about the new instance profile.

Response Lenses

createInstanceProfileResponse_httpStatus :: Lens' CreateInstanceProfileResponse Int

The response's http status code.

createInstanceProfileResponse_instanceProfile :: Lens' CreateInstanceProfileResponse InstanceProfile

A structure containing details about the new instance profile.