Amazonka.IAM.CreateRole

Description

Creates a new role for your Amazon Web Services account. For more information about roles, see IAM roles. For information about quotas for role names and the number of roles you can create, see IAM and STS quotas in the IAM User Guide.

Creating a Request

data CreateRole

See: newCreateRole smart constructor.

Constructors

CreateRole'

Fields

maxSessionDuration :: Maybe Natural

The maximum session duration (in seconds) that you want to set for the specified role. If you do not specify a value for this setting, the default maximum of one hour is applied. This setting can have a value from 1 hour to 12 hours.

Anyone who assumes the role from the or API can use the DurationSeconds API parameter or the duration-seconds CLI parameter to request a longer session. The MaxSessionDuration setting determines the maximum duration that can be requested using the DurationSeconds parameter. If users don't specify a value for the DurationSeconds parameter, their security credentials are valid for one hour by default. This applies when you use the AssumeRole* API operations or the assume-role* CLI operations but does not apply when you use those operations to create a console URL. For more information, see Using IAM roles in the IAM User Guide.

path :: Maybe Text

The path to the role. For more information about paths, see IAM Identifiers in the IAM User Guide.

This parameter is optional. If it is not included, it defaults to a slash (/).

This parameter allows (through its regex pattern) a string of characters consisting of either a forward slash (/) by itself or a string that must begin and end with forward slashes. In addition, it can contain any ASCII character from the ! (\u0021) through the DEL character (\u007F), including most punctuation characters, digits, and upper and lowercased letters.

permissionsBoundary :: Maybe Text

The ARN of the policy that is used to set the permissions boundary for the role.

description :: Maybe Text

A description of the role.

tags :: Maybe [Tag]

A list of tags that you want to attach to the new role. Each tag consists of a key name and an associated value. For more information about tagging, see Tagging IAM resources in the IAM User Guide.

If any one of the tags is invalid or if you exceed the allowed maximum number of tags, then the entire request fails and the resource is not created.

roleName :: Text

The name of the role to create.

IAM user, group, role, and policy names must be unique within the account. Names are not distinguished by case. For example, you cannot create resources named both "MyResource" and "myresource".

assumeRolePolicyDocument :: Text

The trust relationship policy document that grants an entity permission to assume the role.

In IAM, you must provide a JSON policy that has been converted to a string. However, for CloudFormation templates formatted in YAML, you can provide the policy in JSON or YAML format. CloudFormation always converts a YAML policy to JSON format before submitting it to IAM.

The regex pattern used to validate this parameter is a string of characters consisting of the following:

Any printable ASCII character ranging from the space character (\u0020) through the end of the ASCII character range

The printable characters in the Basic Latin and Latin-1 Supplement character set (through \u00FF)

The special characters tab (\u0009), line feed (\u000A), and carriage return (\u000D)

Upon success, the response includes the same trust policy in JSON format.

newCreateRole

Arguments

:: Text
$sel:roleName:CreateRole'

-> Text
$sel:assumeRolePolicyDocument:CreateRole'

-> CreateRole

Create a value of CreateRole with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:maxSessionDuration:CreateRole', createRole_maxSessionDuration - The maximum session duration (in seconds) that you want to set for the specified role. If you do not specify a value for this setting, the default maximum of one hour is applied. This setting can have a value from 1 hour to 12 hours.

Anyone who assumes the role from the or API can use the DurationSeconds API parameter or the duration-seconds CLI parameter to request a longer session. The MaxSessionDuration setting determines the maximum duration that can be requested using the DurationSeconds parameter. If users don't specify a value for the DurationSeconds parameter, their security credentials are valid for one hour by default. This applies when you use the AssumeRole* API operations or the assume-role* CLI operations but does not apply when you use those operations to create a console URL. For more information, see Using IAM roles in the IAM User Guide.

$sel:path:CreateRole', createRole_path - The path to the role. For more information about paths, see IAM Identifiers in the IAM User Guide.

This parameter is optional. If it is not included, it defaults to a slash (/).

This parameter allows (through its regex pattern) a string of characters consisting of either a forward slash (/) by itself or a string that must begin and end with forward slashes. In addition, it can contain any ASCII character from the ! (\u0021) through the DEL character (\u007F), including most punctuation characters, digits, and upper and lowercased letters.

$sel:permissionsBoundary:CreateRole', createRole_permissionsBoundary - The ARN of the policy that is used to set the permissions boundary for the role.

$sel:description:CreateRole', createRole_description - A description of the role.

$sel:tags:CreateRole', createRole_tags - A list of tags that you want to attach to the new role. Each tag consists of a key name and an associated value. For more information about tagging, see Tagging IAM resources in the IAM User Guide.

If any one of the tags is invalid or if you exceed the allowed maximum number of tags, then the entire request fails and the resource is not created.

$sel:roleName:CreateRole', createRole_roleName - The name of the role to create.

IAM user, group, role, and policy names must be unique within the account. Names are not distinguished by case. For example, you cannot create resources named both "MyResource" and "myresource".

$sel:assumeRolePolicyDocument:CreateRole', createRole_assumeRolePolicyDocument - The trust relationship policy document that grants an entity permission to assume the role.

In IAM, you must provide a JSON policy that has been converted to a string. However, for CloudFormation templates formatted in YAML, you can provide the policy in JSON or YAML format. CloudFormation always converts a YAML policy to JSON format before submitting it to IAM.

The regex pattern used to validate this parameter is a string of characters consisting of the following:

Any printable ASCII character ranging from the space character (\u0020) through the end of the ASCII character range

The printable characters in the Basic Latin and Latin-1 Supplement character set (through \u00FF)

The special characters tab (\u0009), line feed (\u000A), and carriage return (\u000D)

Upon success, the response includes the same trust policy in JSON format.

Request Lenses

createRole_maxSessionDuration :: Lens' CreateRole (Maybe Natural)

The maximum session duration (in seconds) that you want to set for the specified role. If you do not specify a value for this setting, the default maximum of one hour is applied. This setting can have a value from 1 hour to 12 hours.

Anyone who assumes the role from the or API can use the DurationSeconds API parameter or the duration-seconds CLI parameter to request a longer session. The MaxSessionDuration setting determines the maximum duration that can be requested using the DurationSeconds parameter. If users don't specify a value for the DurationSeconds parameter, their security credentials are valid for one hour by default. This applies when you use the AssumeRole* API operations or the assume-role* CLI operations but does not apply when you use those operations to create a console URL. For more information, see Using IAM roles in the IAM User Guide.

createRole_path :: Lens' CreateRole (Maybe Text)

The path to the role. For more information about paths, see IAM Identifiers in the IAM User Guide.

This parameter is optional. If it is not included, it defaults to a slash (/).

This parameter allows (through its regex pattern) a string of characters consisting of either a forward slash (/) by itself or a string that must begin and end with forward slashes. In addition, it can contain any ASCII character from the ! (\u0021) through the DEL character (\u007F), including most punctuation characters, digits, and upper and lowercased letters.

createRole_permissionsBoundary :: Lens' CreateRole (Maybe Text)

The ARN of the policy that is used to set the permissions boundary for the role.

createRole_description :: Lens' CreateRole (Maybe Text)

A description of the role.

createRole_tags :: Lens' CreateRole (Maybe [Tag])

A list of tags that you want to attach to the new role. Each tag consists of a key name and an associated value. For more information about tagging, see Tagging IAM resources in the IAM User Guide.

If any one of the tags is invalid or if you exceed the allowed maximum number of tags, then the entire request fails and the resource is not created.

createRole_roleName :: Lens' CreateRole Text

The name of the role to create.

IAM user, group, role, and policy names must be unique within the account. Names are not distinguished by case. For example, you cannot create resources named both "MyResource" and "myresource".

createRole_assumeRolePolicyDocument :: Lens' CreateRole Text

The trust relationship policy document that grants an entity permission to assume the role.

In IAM, you must provide a JSON policy that has been converted to a string. However, for CloudFormation templates formatted in YAML, you can provide the policy in JSON or YAML format. CloudFormation always converts a YAML policy to JSON format before submitting it to IAM.

The regex pattern used to validate this parameter is a string of characters consisting of the following:

Any printable ASCII character ranging from the space character (\u0020) through the end of the ASCII character range

The printable characters in the Basic Latin and Latin-1 Supplement character set (through \u00FF)

The special characters tab (\u0009), line feed (\u000A), and carriage return (\u000D)

Upon success, the response includes the same trust policy in JSON format.

Destructuring the Response

data CreateRoleResponse

Contains the response to a successful CreateRole request.

See: newCreateRoleResponse smart constructor.

Constructors

CreateRoleResponse'

Fields

httpStatus :: Int

The response's http status code.

role' :: Role

A structure containing details about the new role.

newCreateRoleResponse

Arguments

:: Int
$sel:httpStatus:CreateRoleResponse'

-> Role
$sel:role':CreateRoleResponse'

-> CreateRoleResponse

Create a value of CreateRoleResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:httpStatus:CreateRoleResponse', createRoleResponse_httpStatus - The response's http status code.

$sel:role':CreateRoleResponse', createRoleResponse_role - A structure containing details about the new role.

Response Lenses

createRoleResponse_httpStatus :: Lens' CreateRoleResponse Int

The response's http status code.

createRoleResponse_role :: Lens' CreateRoleResponse Role

A structure containing details about the new role.