Amazonka.IAM.DeleteRolePolicy

Description

Deletes the specified inline policy that is embedded in the specified IAM role.

A role can also have managed policies attached to it. To detach a managed policy from a role, use DetachRolePolicy. For more information about policies, refer to Managed policies and inline policies in the IAM User Guide.

Creating a Request

data DeleteRolePolicy

See: newDeleteRolePolicy smart constructor.

Constructors

DeleteRolePolicy'

Fields

roleName :: Text

The name (friendly name, not ARN) identifying the role that the policy is embedded in.

This parameter allows (through its regex pattern) a string of characters consisting of upper and lowercase alphanumeric characters with no spaces. You can also include any of the following characters: _+=,.@-

policyName :: Text

The name of the inline policy to delete from the specified IAM role.

This parameter allows (through its regex pattern) a string of characters consisting of upper and lowercase alphanumeric characters with no spaces. You can also include any of the following characters: _+=,.@-

newDeleteRolePolicy

Arguments

:: Text
$sel:roleName:DeleteRolePolicy'

-> Text
$sel:policyName:DeleteRolePolicy'

-> DeleteRolePolicy

Create a value of DeleteRolePolicy with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:roleName:DeleteRolePolicy', deleteRolePolicy_roleName - The name (friendly name, not ARN) identifying the role that the policy is embedded in.

This parameter allows (through its regex pattern) a string of characters consisting of upper and lowercase alphanumeric characters with no spaces. You can also include any of the following characters: _+=,.@-

$sel:policyName:DeleteRolePolicy', deleteRolePolicy_policyName - The name of the inline policy to delete from the specified IAM role.

This parameter allows (through its regex pattern) a string of characters consisting of upper and lowercase alphanumeric characters with no spaces. You can also include any of the following characters: _+=,.@-

Request Lenses

deleteRolePolicy_roleName :: Lens' DeleteRolePolicy Text

The name (friendly name, not ARN) identifying the role that the policy is embedded in.

This parameter allows (through its regex pattern) a string of characters consisting of upper and lowercase alphanumeric characters with no spaces. You can also include any of the following characters: _+=,.@-

deleteRolePolicy_policyName :: Lens' DeleteRolePolicy Text

The name of the inline policy to delete from the specified IAM role.

This parameter allows (through its regex pattern) a string of characters consisting of upper and lowercase alphanumeric characters with no spaces. You can also include any of the following characters: _+=,.@-

Destructuring the Response

data DeleteRolePolicyResponse

See: newDeleteRolePolicyResponse smart constructor.

Constructors

DeleteRolePolicyResponse'

newDeleteRolePolicyResponse :: DeleteRolePolicyResponse

Create a value of DeleteRolePolicyResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.