Amazonka.IAM.DeleteServerCertificate

Description

Deletes the specified server certificate.

For more information about working with server certificates, see Working with server certificates in the IAM User Guide. This topic also includes a list of Amazon Web Services services that can use the server certificates that you manage with IAM.

If you are using a server certificate with Elastic Load Balancing, deleting the certificate could have implications for your application. If Elastic Load Balancing doesn't detect the deletion of bound certificates, it may continue to use the certificates. This could cause Elastic Load Balancing to stop accepting traffic. We recommend that you remove the reference to the certificate from Elastic Load Balancing before using this command to delete the certificate. For more information, see DeleteLoadBalancerListeners in the Elastic Load Balancing API Reference.

Creating a Request

data DeleteServerCertificate

See: newDeleteServerCertificate smart constructor.

Constructors

DeleteServerCertificate'

Fields

serverCertificateName :: Text

The name of the server certificate you want to delete.

This parameter allows (through its regex pattern) a string of characters consisting of upper and lowercase alphanumeric characters with no spaces. You can also include any of the following characters: _+=,.@-

newDeleteServerCertificate

Arguments

:: Text
$sel:serverCertificateName:DeleteServerCertificate'

-> DeleteServerCertificate

Create a value of DeleteServerCertificate with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:serverCertificateName:DeleteServerCertificate', deleteServerCertificate_serverCertificateName - The name of the server certificate you want to delete.

This parameter allows (through its regex pattern) a string of characters consisting of upper and lowercase alphanumeric characters with no spaces. You can also include any of the following characters: _+=,.@-

Request Lenses

deleteServerCertificate_serverCertificateName :: Lens' DeleteServerCertificate Text

The name of the server certificate you want to delete.

This parameter allows (through its regex pattern) a string of characters consisting of upper and lowercase alphanumeric characters with no spaces. You can also include any of the following characters: _+=,.@-

Destructuring the Response

data DeleteServerCertificateResponse

See: newDeleteServerCertificateResponse smart constructor.

Constructors

DeleteServerCertificateResponse'

newDeleteServerCertificateResponse :: DeleteServerCertificateResponse

Create a value of DeleteServerCertificateResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.