Amazonka.IAM.DeleteServiceLinkedRole

Description

Submits a service-linked role deletion request and returns a DeletionTaskId, which you can use to check the status of the deletion. Before you call this operation, confirm that the role has no active sessions and that any resources used by the role in the linked service are deleted. If you call this operation more than once for the same service-linked role and an earlier deletion task is not complete, then the DeletionTaskId of the earlier request is returned.

If you submit a deletion request for a service-linked role whose linked service is still accessing a resource, then the deletion task fails. If it fails, the GetServiceLinkedRoleDeletionStatus operation returns the reason for the failure, usually including the resources that must be deleted. To delete the service-linked role, you must first remove those resources from the linked service and then submit the deletion request again. Resources are specific to the service that is linked to the role. For more information about removing resources from a service, see the Amazon Web Services documentation for your service.

For more information about service-linked roles, see Roles terms and concepts: Amazon Web Services service-linked role in the IAM User Guide.

Creating a Request

data DeleteServiceLinkedRole

See: newDeleteServiceLinkedRole smart constructor.

Constructors

DeleteServiceLinkedRole'

Fields

roleName :: Text

The name of the service-linked role to be deleted.

newDeleteServiceLinkedRole

Arguments

:: Text
$sel:roleName:DeleteServiceLinkedRole'

-> DeleteServiceLinkedRole

Create a value of DeleteServiceLinkedRole with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:roleName:DeleteServiceLinkedRole', deleteServiceLinkedRole_roleName - The name of the service-linked role to be deleted.

Request Lenses

deleteServiceLinkedRole_roleName :: Lens' DeleteServiceLinkedRole Text

The name of the service-linked role to be deleted.

Destructuring the Response

data DeleteServiceLinkedRoleResponse

See: newDeleteServiceLinkedRoleResponse smart constructor.

Constructors

DeleteServiceLinkedRoleResponse'

Fields

httpStatus :: Int

The response's http status code.

deletionTaskId :: Text

The deletion task identifier that you can use to check the status of the deletion. This identifier is returned in the format task/aws-service-role/<service-principal-name>/<role-name>/<task-uuid>.

newDeleteServiceLinkedRoleResponse

Arguments

:: Int
$sel:httpStatus:DeleteServiceLinkedRoleResponse'

-> Text
$sel:deletionTaskId:DeleteServiceLinkedRoleResponse'

-> DeleteServiceLinkedRoleResponse

Create a value of DeleteServiceLinkedRoleResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:httpStatus:DeleteServiceLinkedRoleResponse', deleteServiceLinkedRoleResponse_httpStatus - The response's http status code.

$sel:deletionTaskId:DeleteServiceLinkedRoleResponse', deleteServiceLinkedRoleResponse_deletionTaskId - The deletion task identifier that you can use to check the status of the deletion. This identifier is returned in the format task/aws-service-role/<service-principal-name>/<role-name>/<task-uuid>.

Response Lenses

deleteServiceLinkedRoleResponse_httpStatus :: Lens' DeleteServiceLinkedRoleResponse Int

The response's http status code.

deleteServiceLinkedRoleResponse_deletionTaskId :: Lens' DeleteServiceLinkedRoleResponse Text

The deletion task identifier that you can use to check the status of the deletion. This identifier is returned in the format task/aws-service-role/<service-principal-name>/<role-name>/<task-uuid>.