Amazonka.IAM.DeleteUserPolicy

Description

Deletes the specified inline policy that is embedded in the specified IAM user.

A user can also have managed policies attached to it. To detach a managed policy from a user, use DetachUserPolicy. For more information about policies, refer to Managed policies and inline policies in the IAM User Guide.

Creating a Request

data DeleteUserPolicy

See: newDeleteUserPolicy smart constructor.

Constructors

DeleteUserPolicy'

Fields

userName :: Text

The name (friendly name, not ARN) identifying the user that the policy is embedded in.

This parameter allows (through its regex pattern) a string of characters consisting of upper and lowercase alphanumeric characters with no spaces. You can also include any of the following characters: _+=,.@-

policyName :: Text

The name identifying the policy document to delete.

This parameter allows (through its regex pattern) a string of characters consisting of upper and lowercase alphanumeric characters with no spaces. You can also include any of the following characters: _+=,.@-

newDeleteUserPolicy

Arguments

:: Text
$sel:userName:DeleteUserPolicy'

-> Text
$sel:policyName:DeleteUserPolicy'

-> DeleteUserPolicy

Create a value of DeleteUserPolicy with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:userName:DeleteUserPolicy', deleteUserPolicy_userName - The name (friendly name, not ARN) identifying the user that the policy is embedded in.

This parameter allows (through its regex pattern) a string of characters consisting of upper and lowercase alphanumeric characters with no spaces. You can also include any of the following characters: _+=,.@-

$sel:policyName:DeleteUserPolicy', deleteUserPolicy_policyName - The name identifying the policy document to delete.

This parameter allows (through its regex pattern) a string of characters consisting of upper and lowercase alphanumeric characters with no spaces. You can also include any of the following characters: _+=,.@-

Request Lenses

deleteUserPolicy_userName :: Lens' DeleteUserPolicy Text

The name (friendly name, not ARN) identifying the user that the policy is embedded in.

This parameter allows (through its regex pattern) a string of characters consisting of upper and lowercase alphanumeric characters with no spaces. You can also include any of the following characters: _+=,.@-

deleteUserPolicy_policyName :: Lens' DeleteUserPolicy Text

The name identifying the policy document to delete.

This parameter allows (through its regex pattern) a string of characters consisting of upper and lowercase alphanumeric characters with no spaces. You can also include any of the following characters: _+=,.@-

Destructuring the Response

data DeleteUserPolicyResponse

See: newDeleteUserPolicyResponse smart constructor.

Constructors

DeleteUserPolicyResponse'

newDeleteUserPolicyResponse :: DeleteUserPolicyResponse

Create a value of DeleteUserPolicyResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.