Amazonka.IAM.GetContextKeysForCustomPolicy

Description

Gets a list of all of the context keys referenced in the input policies. The policies are supplied as a list of one or more strings. To get the context keys from policies associated with an IAM user, group, or role, use GetContextKeysForPrincipalPolicy.

Context keys are variables maintained by Amazon Web Services and its services that provide details about the context of an API query request. Context keys can be evaluated by testing against a value specified in an IAM policy. Use GetContextKeysForCustomPolicy to understand what key names and values you must supply when you call SimulateCustomPolicy. Note that all parameters are shown in unencoded form here for clarity but must be URL encoded to be included as a part of a real HTML request.

Creating a Request

data GetContextKeysForCustomPolicy

See: newGetContextKeysForCustomPolicy smart constructor.

Constructors

GetContextKeysForCustomPolicy'

Fields

policyInputList :: [Text]

A list of policies for which you want the list of context keys referenced in those policies. Each document is specified as a string containing the complete, valid JSON text of an IAM policy.

The regex pattern used to validate this parameter is a string of characters consisting of the following:

Any printable ASCII character ranging from the space character (\u0020) through the end of the ASCII character range

The printable characters in the Basic Latin and Latin-1 Supplement character set (through \u00FF)

The special characters tab (\u0009), line feed (\u000A), and carriage return (\u000D)

newGetContextKeysForCustomPolicy :: GetContextKeysForCustomPolicy

Create a value of GetContextKeysForCustomPolicy with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:policyInputList:GetContextKeysForCustomPolicy', getContextKeysForCustomPolicy_policyInputList - A list of policies for which you want the list of context keys referenced in those policies. Each document is specified as a string containing the complete, valid JSON text of an IAM policy.

The regex pattern used to validate this parameter is a string of characters consisting of the following:

Any printable ASCII character ranging from the space character (\u0020) through the end of the ASCII character range

The printable characters in the Basic Latin and Latin-1 Supplement character set (through \u00FF)

The special characters tab (\u0009), line feed (\u000A), and carriage return (\u000D)

Request Lenses

getContextKeysForCustomPolicy_policyInputList :: Lens' GetContextKeysForCustomPolicy [Text]

A list of policies for which you want the list of context keys referenced in those policies. Each document is specified as a string containing the complete, valid JSON text of an IAM policy.

The regex pattern used to validate this parameter is a string of characters consisting of the following:

Any printable ASCII character ranging from the space character (\u0020) through the end of the ASCII character range

The printable characters in the Basic Latin and Latin-1 Supplement character set (through \u00FF)

The special characters tab (\u0009), line feed (\u000A), and carriage return (\u000D)

Destructuring the Response

data GetContextKeysForPolicyResponse

Contains the response to a successful GetContextKeysForPrincipalPolicy or GetContextKeysForCustomPolicy request.

See: newGetContextKeysForPolicyResponse smart constructor.

Constructors

GetContextKeysForPolicyResponse'

Fields

contextKeyNames :: Maybe [Text]

The list of context keys that are referenced in the input policies.

newGetContextKeysForPolicyResponse :: GetContextKeysForPolicyResponse

Create a value of GetContextKeysForPolicyResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:contextKeyNames:GetContextKeysForPolicyResponse', getContextKeysForPolicyResponse_contextKeyNames - The list of context keys that are referenced in the input policies.

Response Lenses

getContextKeysForPolicyResponse_contextKeyNames :: Lens' GetContextKeysForPolicyResponse (Maybe [Text])

The list of context keys that are referenced in the input policies.