Amazonka.IAM.GetLoginProfile

Description

Retrieves the user name for the specified IAM user. A login profile is created when you create a password for the user to access the Amazon Web Services Management Console. If the user does not exist or does not have a password, the operation returns a 404 (NoSuchEntity) error.

If you create an IAM user with access to the console, the CreateDate reflects the date you created the initial password for the user.

If you create an IAM user with programmatic access, and then later add a password for the user to access the Amazon Web Services Management Console, the CreateDate reflects the initial password creation date. A user with programmatic access does not have a login profile unless you create a password for the user to access the Amazon Web Services Management Console.

Creating a Request

data GetLoginProfile

See: newGetLoginProfile smart constructor.

Constructors

GetLoginProfile'

Fields

userName :: Text

The name of the user whose login profile you want to retrieve.

This parameter allows (through its regex pattern) a string of characters consisting of upper and lowercase alphanumeric characters with no spaces. You can also include any of the following characters: _+=,.@-

newGetLoginProfile

Arguments

:: Text
$sel:userName:GetLoginProfile'

-> GetLoginProfile

Create a value of GetLoginProfile with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:userName:GetLoginProfile', getLoginProfile_userName - The name of the user whose login profile you want to retrieve.

This parameter allows (through its regex pattern) a string of characters consisting of upper and lowercase alphanumeric characters with no spaces. You can also include any of the following characters: _+=,.@-

Request Lenses

getLoginProfile_userName :: Lens' GetLoginProfile Text

The name of the user whose login profile you want to retrieve.

This parameter allows (through its regex pattern) a string of characters consisting of upper and lowercase alphanumeric characters with no spaces. You can also include any of the following characters: _+=,.@-

Destructuring the Response

data GetLoginProfileResponse

Contains the response to a successful GetLoginProfile request.

See: newGetLoginProfileResponse smart constructor.

Constructors

GetLoginProfileResponse'

Fields

httpStatus :: Int

The response's http status code.

loginProfile :: LoginProfile

A structure containing the user name and the profile creation date for the user.

newGetLoginProfileResponse

Arguments

:: Int
$sel:httpStatus:GetLoginProfileResponse'

-> LoginProfile
$sel:loginProfile:GetLoginProfileResponse'

-> GetLoginProfileResponse

Create a value of GetLoginProfileResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:httpStatus:GetLoginProfileResponse', getLoginProfileResponse_httpStatus - The response's http status code.

$sel:loginProfile:GetLoginProfileResponse', getLoginProfileResponse_loginProfile - A structure containing the user name and the profile creation date for the user.

Response Lenses

getLoginProfileResponse_httpStatus :: Lens' GetLoginProfileResponse Int

The response's http status code.

getLoginProfileResponse_loginProfile :: Lens' GetLoginProfileResponse LoginProfile

A structure containing the user name and the profile creation date for the user.