Amazonka.IAM.GetRolePolicy

Description

Retrieves the specified inline policy document that is embedded with the specified IAM role.

Policies returned by this operation are URL-encoded compliant with RFC 3986. You can use a URL decoding method to convert the policy back to plain JSON text. For example, if you use Java, you can use the decode method of the java.net.URLDecoder utility class in the Java SDK. Other languages and SDKs provide similar functionality.

An IAM role can also have managed policies attached to it. To retrieve a managed policy document that is attached to a role, use GetPolicy to determine the policy's default version, then use GetPolicyVersion to retrieve the policy document.

For more information about policies, see Managed policies and inline policies in the IAM User Guide.

For more information about roles, see Using roles to delegate permissions and federate identities.

Creating a Request

data GetRolePolicy

See: newGetRolePolicy smart constructor.

Constructors

GetRolePolicy'

Fields

roleName :: Text

The name of the role associated with the policy.

This parameter allows (through its regex pattern) a string of characters consisting of upper and lowercase alphanumeric characters with no spaces. You can also include any of the following characters: _+=,.@-

policyName :: Text

The name of the policy document to get.

This parameter allows (through its regex pattern) a string of characters consisting of upper and lowercase alphanumeric characters with no spaces. You can also include any of the following characters: _+=,.@-

newGetRolePolicy

Arguments

:: Text
$sel:roleName:GetRolePolicy'

-> Text
$sel:policyName:GetRolePolicy'

-> GetRolePolicy

Create a value of GetRolePolicy with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:roleName:GetRolePolicy', getRolePolicy_roleName - The name of the role associated with the policy.

This parameter allows (through its regex pattern) a string of characters consisting of upper and lowercase alphanumeric characters with no spaces. You can also include any of the following characters: _+=,.@-

$sel:policyName:GetRolePolicy', getRolePolicy_policyName - The name of the policy document to get.

This parameter allows (through its regex pattern) a string of characters consisting of upper and lowercase alphanumeric characters with no spaces. You can also include any of the following characters: _+=,.@-

Request Lenses

getRolePolicy_roleName :: Lens' GetRolePolicy Text

The name of the role associated with the policy.

This parameter allows (through its regex pattern) a string of characters consisting of upper and lowercase alphanumeric characters with no spaces. You can also include any of the following characters: _+=,.@-

getRolePolicy_policyName :: Lens' GetRolePolicy Text

The name of the policy document to get.

This parameter allows (through its regex pattern) a string of characters consisting of upper and lowercase alphanumeric characters with no spaces. You can also include any of the following characters: _+=,.@-

Destructuring the Response

data GetRolePolicyResponse

Contains the response to a successful GetRolePolicy request.

See: newGetRolePolicyResponse smart constructor.

Constructors

GetRolePolicyResponse'

Fields

httpStatus :: Int

The response's http status code.

roleName :: Text

The role the policy is associated with.

policyName :: Text

The name of the policy.

policyDocument :: Text

The policy document.

IAM stores policies in JSON format. However, resources that were created using CloudFormation templates can be formatted in YAML. CloudFormation always converts a YAML policy to JSON format before submitting it to IAM.

newGetRolePolicyResponse

Arguments

:: Int
$sel:httpStatus:GetRolePolicyResponse'

-> Text
$sel:roleName:GetRolePolicy'

-> Text
$sel:policyName:GetRolePolicy'

-> Text
$sel:policyDocument:GetRolePolicyResponse'

-> GetRolePolicyResponse

Create a value of GetRolePolicyResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:httpStatus:GetRolePolicyResponse', getRolePolicyResponse_httpStatus - The response's http status code.

$sel:roleName:GetRolePolicy', getRolePolicyResponse_roleName - The role the policy is associated with.

$sel:policyName:GetRolePolicy', getRolePolicyResponse_policyName - The name of the policy.

$sel:policyDocument:GetRolePolicyResponse', getRolePolicyResponse_policyDocument - The policy document.

IAM stores policies in JSON format. However, resources that were created using CloudFormation templates can be formatted in YAML. CloudFormation always converts a YAML policy to JSON format before submitting it to IAM.

Response Lenses

getRolePolicyResponse_httpStatus :: Lens' GetRolePolicyResponse Int

The response's http status code.

getRolePolicyResponse_roleName :: Lens' GetRolePolicyResponse Text

The role the policy is associated with.

getRolePolicyResponse_policyName :: Lens' GetRolePolicyResponse Text

The name of the policy.

getRolePolicyResponse_policyDocument :: Lens' GetRolePolicyResponse Text

The policy document.

IAM stores policies in JSON format. However, resources that were created using CloudFormation templates can be formatted in YAML. CloudFormation always converts a YAML policy to JSON format before submitting it to IAM.