Amazonka.IAM.ListInstanceProfileTags

Description

Lists the tags that are attached to the specified IAM instance profile. The returned list of tags is sorted by tag key. For more information about tagging, see Tagging IAM resources in the IAM User Guide.

Creating a Request

data ListInstanceProfileTags

See: newListInstanceProfileTags smart constructor.

Constructors

ListInstanceProfileTags'

Fields

marker :: Maybe Text

Use this parameter only when paginating results and only after you receive a response indicating that the results are truncated. Set it to the value of the Marker element in the response that you received to indicate where the next call should start.

maxItems :: Maybe Natural

Use this only when paginating results to indicate the maximum number of items you want in the response. If additional items exist beyond the maximum you specify, the IsTruncated response element is true.

If you do not include this parameter, the number of items defaults to 100. Note that IAM might return fewer results, even when there are more results available. In that case, the IsTruncated response element returns true, and Marker contains a value to include in the subsequent call that tells the service where to continue from.

instanceProfileName :: Text

The name of the IAM instance profile whose tags you want to see.

This parameter allows (through its regex pattern) a string of characters consisting of upper and lowercase alphanumeric characters with no spaces. You can also include any of the following characters: _+=,.@-

newListInstanceProfileTags

Arguments

:: Text
$sel:instanceProfileName:ListInstanceProfileTags'

-> ListInstanceProfileTags

Create a value of ListInstanceProfileTags with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:marker:ListInstanceProfileTags', listInstanceProfileTags_marker - Use this parameter only when paginating results and only after you receive a response indicating that the results are truncated. Set it to the value of the Marker element in the response that you received to indicate where the next call should start.

$sel:maxItems:ListInstanceProfileTags', listInstanceProfileTags_maxItems - Use this only when paginating results to indicate the maximum number of items you want in the response. If additional items exist beyond the maximum you specify, the IsTruncated response element is true.

If you do not include this parameter, the number of items defaults to 100. Note that IAM might return fewer results, even when there are more results available. In that case, the IsTruncated response element returns true, and Marker contains a value to include in the subsequent call that tells the service where to continue from.

$sel:instanceProfileName:ListInstanceProfileTags', listInstanceProfileTags_instanceProfileName - The name of the IAM instance profile whose tags you want to see.

This parameter allows (through its regex pattern) a string of characters consisting of upper and lowercase alphanumeric characters with no spaces. You can also include any of the following characters: _+=,.@-

Request Lenses

listInstanceProfileTags_marker :: Lens' ListInstanceProfileTags (Maybe Text)

Use this parameter only when paginating results and only after you receive a response indicating that the results are truncated. Set it to the value of the Marker element in the response that you received to indicate where the next call should start.

listInstanceProfileTags_maxItems :: Lens' ListInstanceProfileTags (Maybe Natural)

Use this only when paginating results to indicate the maximum number of items you want in the response. If additional items exist beyond the maximum you specify, the IsTruncated response element is true.

If you do not include this parameter, the number of items defaults to 100. Note that IAM might return fewer results, even when there are more results available. In that case, the IsTruncated response element returns true, and Marker contains a value to include in the subsequent call that tells the service where to continue from.

listInstanceProfileTags_instanceProfileName :: Lens' ListInstanceProfileTags Text

The name of the IAM instance profile whose tags you want to see.

This parameter allows (through its regex pattern) a string of characters consisting of upper and lowercase alphanumeric characters with no spaces. You can also include any of the following characters: _+=,.@-

Destructuring the Response

data ListInstanceProfileTagsResponse

See: newListInstanceProfileTagsResponse smart constructor.

Constructors

ListInstanceProfileTagsResponse'

Fields

marker :: Maybe Text

When IsTruncated is true, this element is present and contains the value to use for the Marker parameter in a subsequent pagination request.

isTruncated :: Maybe Bool

A flag that indicates whether there are more items to return. If your results were truncated, you can make a subsequent pagination request using the Marker request parameter to retrieve more items. Note that IAM might return fewer than the MaxItems number of results even when there are more results available. We recommend that you check IsTruncated after every call to ensure that you receive all your results.

httpStatus :: Int

The response's http status code.

tags :: [Tag]

The list of tags that are currently attached to the IAM instance profile. Each tag consists of a key name and an associated value. If no tags are attached to the specified resource, the response contains an empty list.

newListInstanceProfileTagsResponse

Arguments

:: Int
$sel:httpStatus:ListInstanceProfileTagsResponse'

-> ListInstanceProfileTagsResponse

Create a value of ListInstanceProfileTagsResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:marker:ListInstanceProfileTags', listInstanceProfileTagsResponse_marker - When IsTruncated is true, this element is present and contains the value to use for the Marker parameter in a subsequent pagination request.

$sel:isTruncated:ListInstanceProfileTagsResponse', listInstanceProfileTagsResponse_isTruncated - A flag that indicates whether there are more items to return. If your results were truncated, you can make a subsequent pagination request using the Marker request parameter to retrieve more items. Note that IAM might return fewer than the MaxItems number of results even when there are more results available. We recommend that you check IsTruncated after every call to ensure that you receive all your results.

$sel:httpStatus:ListInstanceProfileTagsResponse', listInstanceProfileTagsResponse_httpStatus - The response's http status code.

$sel:tags:ListInstanceProfileTagsResponse', listInstanceProfileTagsResponse_tags - The list of tags that are currently attached to the IAM instance profile. Each tag consists of a key name and an associated value. If no tags are attached to the specified resource, the response contains an empty list.

Response Lenses

listInstanceProfileTagsResponse_marker :: Lens' ListInstanceProfileTagsResponse (Maybe Text)

When IsTruncated is true, this element is present and contains the value to use for the Marker parameter in a subsequent pagination request.

listInstanceProfileTagsResponse_isTruncated :: Lens' ListInstanceProfileTagsResponse (Maybe Bool)

A flag that indicates whether there are more items to return. If your results were truncated, you can make a subsequent pagination request using the Marker request parameter to retrieve more items. Note that IAM might return fewer than the MaxItems number of results even when there are more results available. We recommend that you check IsTruncated after every call to ensure that you receive all your results.

listInstanceProfileTagsResponse_httpStatus :: Lens' ListInstanceProfileTagsResponse Int

The response's http status code.

listInstanceProfileTagsResponse_tags :: Lens' ListInstanceProfileTagsResponse [Tag]

The list of tags that are currently attached to the IAM instance profile. Each tag consists of a key name and an associated value. If no tags are attached to the specified resource, the response contains an empty list.