Amazonka.IAM.ListSSHPublicKeys

Description

Returns information about the SSH public keys associated with the specified IAM user. If none exists, the operation returns an empty list.

The SSH public keys returned by this operation are used only for authenticating the IAM user to an CodeCommit repository. For more information about using SSH keys to authenticate to an CodeCommit repository, see Set up CodeCommit for SSH connections in the CodeCommit User Guide.

Although each user is limited to a small number of keys, you can still paginate the results using the MaxItems and Marker parameters.

This operation returns paginated results.

Creating a Request

data ListSSHPublicKeys

See: newListSSHPublicKeys smart constructor.

Constructors

ListSSHPublicKeys'

Fields

userName :: Maybe Text

The name of the IAM user to list SSH public keys for. If none is specified, the UserName field is determined implicitly based on the Amazon Web Services access key used to sign the request.

This parameter allows (through its regex pattern) a string of characters consisting of upper and lowercase alphanumeric characters with no spaces. You can also include any of the following characters: _+=,.@-

marker :: Maybe Text

Use this parameter only when paginating results and only after you receive a response indicating that the results are truncated. Set it to the value of the Marker element in the response that you received to indicate where the next call should start.

maxItems :: Maybe Natural

Use this only when paginating results to indicate the maximum number of items you want in the response. If additional items exist beyond the maximum you specify, the IsTruncated response element is true.

If you do not include this parameter, the number of items defaults to 100. Note that IAM might return fewer results, even when there are more results available. In that case, the IsTruncated response element returns true, and Marker contains a value to include in the subsequent call that tells the service where to continue from.

newListSSHPublicKeys :: ListSSHPublicKeys

Create a value of ListSSHPublicKeys with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:userName:ListSSHPublicKeys', listSSHPublicKeys_userName - The name of the IAM user to list SSH public keys for. If none is specified, the UserName field is determined implicitly based on the Amazon Web Services access key used to sign the request.

This parameter allows (through its regex pattern) a string of characters consisting of upper and lowercase alphanumeric characters with no spaces. You can also include any of the following characters: _+=,.@-

$sel:marker:ListSSHPublicKeys', listSSHPublicKeys_marker - Use this parameter only when paginating results and only after you receive a response indicating that the results are truncated. Set it to the value of the Marker element in the response that you received to indicate where the next call should start.

$sel:maxItems:ListSSHPublicKeys', listSSHPublicKeys_maxItems - Use this only when paginating results to indicate the maximum number of items you want in the response. If additional items exist beyond the maximum you specify, the IsTruncated response element is true.

If you do not include this parameter, the number of items defaults to 100. Note that IAM might return fewer results, even when there are more results available. In that case, the IsTruncated response element returns true, and Marker contains a value to include in the subsequent call that tells the service where to continue from.

Request Lenses

listSSHPublicKeys_userName :: Lens' ListSSHPublicKeys (Maybe Text)

The name of the IAM user to list SSH public keys for. If none is specified, the UserName field is determined implicitly based on the Amazon Web Services access key used to sign the request.

This parameter allows (through its regex pattern) a string of characters consisting of upper and lowercase alphanumeric characters with no spaces. You can also include any of the following characters: _+=,.@-

listSSHPublicKeys_marker :: Lens' ListSSHPublicKeys (Maybe Text)

Use this parameter only when paginating results and only after you receive a response indicating that the results are truncated. Set it to the value of the Marker element in the response that you received to indicate where the next call should start.

listSSHPublicKeys_maxItems :: Lens' ListSSHPublicKeys (Maybe Natural)

Use this only when paginating results to indicate the maximum number of items you want in the response. If additional items exist beyond the maximum you specify, the IsTruncated response element is true.

If you do not include this parameter, the number of items defaults to 100. Note that IAM might return fewer results, even when there are more results available. In that case, the IsTruncated response element returns true, and Marker contains a value to include in the subsequent call that tells the service where to continue from.

Destructuring the Response

data ListSSHPublicKeysResponse

Contains the response to a successful ListSSHPublicKeys request.

See: newListSSHPublicKeysResponse smart constructor.

Constructors

ListSSHPublicKeysResponse'

Fields

sSHPublicKeys :: Maybe [SSHPublicKeyMetadata]

A list of the SSH public keys assigned to IAM user.

marker :: Maybe Text

When IsTruncated is true, this element is present and contains the value to use for the Marker parameter in a subsequent pagination request.

isTruncated :: Maybe Bool

A flag that indicates whether there are more items to return. If your results were truncated, you can make a subsequent pagination request using the Marker request parameter to retrieve more items. Note that IAM might return fewer than the MaxItems number of results even when there are more results available. We recommend that you check IsTruncated after every call to ensure that you receive all your results.

httpStatus :: Int

The response's http status code.

newListSSHPublicKeysResponse

Arguments

:: Int
$sel:httpStatus:ListSSHPublicKeysResponse'

-> ListSSHPublicKeysResponse

Create a value of ListSSHPublicKeysResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:sSHPublicKeys:ListSSHPublicKeysResponse', listSSHPublicKeysResponse_sSHPublicKeys - A list of the SSH public keys assigned to IAM user.

$sel:marker:ListSSHPublicKeys', listSSHPublicKeysResponse_marker - When IsTruncated is true, this element is present and contains the value to use for the Marker parameter in a subsequent pagination request.

$sel:isTruncated:ListSSHPublicKeysResponse', listSSHPublicKeysResponse_isTruncated - A flag that indicates whether there are more items to return. If your results were truncated, you can make a subsequent pagination request using the Marker request parameter to retrieve more items. Note that IAM might return fewer than the MaxItems number of results even when there are more results available. We recommend that you check IsTruncated after every call to ensure that you receive all your results.

$sel:httpStatus:ListSSHPublicKeysResponse', listSSHPublicKeysResponse_httpStatus - The response's http status code.

Response Lenses

listSSHPublicKeysResponse_sSHPublicKeys :: Lens' ListSSHPublicKeysResponse (Maybe [SSHPublicKeyMetadata])

A list of the SSH public keys assigned to IAM user.

listSSHPublicKeysResponse_marker :: Lens' ListSSHPublicKeysResponse (Maybe Text)

When IsTruncated is true, this element is present and contains the value to use for the Marker parameter in a subsequent pagination request.

listSSHPublicKeysResponse_isTruncated :: Lens' ListSSHPublicKeysResponse (Maybe Bool)

A flag that indicates whether there are more items to return. If your results were truncated, you can make a subsequent pagination request using the Marker request parameter to retrieve more items. Note that IAM might return fewer than the MaxItems number of results even when there are more results available. We recommend that you check IsTruncated after every call to ensure that you receive all your results.

listSSHPublicKeysResponse_httpStatus :: Lens' ListSSHPublicKeysResponse Int

The response's http status code.