Amazonka.IAM.ListServiceSpecificCredentials

Description

Returns information about the service-specific credentials associated with the specified IAM user. If none exists, the operation returns an empty list. The service-specific credentials returned by this operation are used only for authenticating the IAM user to a specific service. For more information about using service-specific credentials to authenticate to an Amazon Web Services service, see Set up service-specific credentials in the CodeCommit User Guide.

Creating a Request

data ListServiceSpecificCredentials

See: newListServiceSpecificCredentials smart constructor.

Constructors

ListServiceSpecificCredentials'

Fields

userName :: Maybe Text

The name of the user whose service-specific credentials you want information about. If this value is not specified, then the operation assumes the user whose credentials are used to call the operation.

This parameter allows (through its regex pattern) a string of characters consisting of upper and lowercase alphanumeric characters with no spaces. You can also include any of the following characters: _+=,.@-

serviceName :: Maybe Text

Filters the returned results to only those for the specified Amazon Web Services service. If not specified, then Amazon Web Services returns service-specific credentials for all services.

newListServiceSpecificCredentials :: ListServiceSpecificCredentials

Create a value of ListServiceSpecificCredentials with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:userName:ListServiceSpecificCredentials', listServiceSpecificCredentials_userName - The name of the user whose service-specific credentials you want information about. If this value is not specified, then the operation assumes the user whose credentials are used to call the operation.

This parameter allows (through its regex pattern) a string of characters consisting of upper and lowercase alphanumeric characters with no spaces. You can also include any of the following characters: _+=,.@-

$sel:serviceName:ListServiceSpecificCredentials', listServiceSpecificCredentials_serviceName - Filters the returned results to only those for the specified Amazon Web Services service. If not specified, then Amazon Web Services returns service-specific credentials for all services.

Request Lenses

listServiceSpecificCredentials_userName :: Lens' ListServiceSpecificCredentials (Maybe Text)

The name of the user whose service-specific credentials you want information about. If this value is not specified, then the operation assumes the user whose credentials are used to call the operation.

This parameter allows (through its regex pattern) a string of characters consisting of upper and lowercase alphanumeric characters with no spaces. You can also include any of the following characters: _+=,.@-

listServiceSpecificCredentials_serviceName :: Lens' ListServiceSpecificCredentials (Maybe Text)

Filters the returned results to only those for the specified Amazon Web Services service. If not specified, then Amazon Web Services returns service-specific credentials for all services.

Destructuring the Response

data ListServiceSpecificCredentialsResponse

See: newListServiceSpecificCredentialsResponse smart constructor.

Constructors

ListServiceSpecificCredentialsResponse'

Fields

serviceSpecificCredentials :: Maybe [ServiceSpecificCredentialMetadata]

A list of structures that each contain details about a service-specific credential.

httpStatus :: Int

The response's http status code.

newListServiceSpecificCredentialsResponse

Arguments

:: Int
$sel:httpStatus:ListServiceSpecificCredentialsResponse'

-> ListServiceSpecificCredentialsResponse

Create a value of ListServiceSpecificCredentialsResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:serviceSpecificCredentials:ListServiceSpecificCredentialsResponse', listServiceSpecificCredentialsResponse_serviceSpecificCredentials - A list of structures that each contain details about a service-specific credential.

$sel:httpStatus:ListServiceSpecificCredentialsResponse', listServiceSpecificCredentialsResponse_httpStatus - The response's http status code.

Response Lenses

listServiceSpecificCredentialsResponse_serviceSpecificCredentials :: Lens' ListServiceSpecificCredentialsResponse (Maybe [ServiceSpecificCredentialMetadata])

A list of structures that each contain details about a service-specific credential.

listServiceSpecificCredentialsResponse_httpStatus :: Lens' ListServiceSpecificCredentialsResponse Int

The response's http status code.