Amazonka.IAM.TagMFADevice

Description

Adds one or more tags to an IAM virtual multi-factor authentication (MFA) device. If a tag with the same key name already exists, then that tag is overwritten with the new value.

A tag consists of a key name and an associated value. By assigning tags to your resources, you can do the following:

Administrative grouping and discovery - Attach tags to resources to aid in organization and search. For example, you could search for all resources with the key name Project and the value MyImportantProject. Or search for all resources with the key name Cost Center and the value 41200.

Access control - Include tags in IAM user-based and resource-based policies. You can use tags to restrict access to only an IAM virtual MFA device that has a specified tag attached. For examples of policies that show how to use tags to control access, see Control access using IAM tags in the IAM User Guide.

If any one of the tags is invalid or if you exceed the allowed maximum number of tags, then the entire request fails and the resource is not created. For more information about tagging, see Tagging IAM resources in the IAM User Guide.

Amazon Web Services always interprets the tag Value as a single string. If you need to store an array, you can store comma-separated values in the string. However, you must interpret the value in your code.

Creating a Request

data TagMFADevice

See: newTagMFADevice smart constructor.

Constructors

TagMFADevice'

Fields

serialNumber :: Text

The unique identifier for the IAM virtual MFA device to which you want to add tags. For virtual MFA devices, the serial number is the same as the ARN.

This parameter allows (through its regex pattern) a string of characters consisting of upper and lowercase alphanumeric characters with no spaces. You can also include any of the following characters: _+=,.@-

tags :: [Tag]

The list of tags that you want to attach to the IAM virtual MFA device. Each tag consists of a key name and an associated value.

newTagMFADevice

Arguments

:: Text
$sel:serialNumber:TagMFADevice'

-> TagMFADevice

Create a value of TagMFADevice with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:serialNumber:TagMFADevice', tagMFADevice_serialNumber - The unique identifier for the IAM virtual MFA device to which you want to add tags. For virtual MFA devices, the serial number is the same as the ARN.

This parameter allows (through its regex pattern) a string of characters consisting of upper and lowercase alphanumeric characters with no spaces. You can also include any of the following characters: _+=,.@-

$sel:tags:TagMFADevice', tagMFADevice_tags - The list of tags that you want to attach to the IAM virtual MFA device. Each tag consists of a key name and an associated value.

Request Lenses

tagMFADevice_serialNumber :: Lens' TagMFADevice Text

The unique identifier for the IAM virtual MFA device to which you want to add tags. For virtual MFA devices, the serial number is the same as the ARN.

This parameter allows (through its regex pattern) a string of characters consisting of upper and lowercase alphanumeric characters with no spaces. You can also include any of the following characters: _+=,.@-

tagMFADevice_tags :: Lens' TagMFADevice [Tag]

The list of tags that you want to attach to the IAM virtual MFA device. Each tag consists of a key name and an associated value.

Destructuring the Response

data TagMFADeviceResponse

See: newTagMFADeviceResponse smart constructor.

Constructors

TagMFADeviceResponse'

newTagMFADeviceResponse :: TagMFADeviceResponse

Create a value of TagMFADeviceResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.