Amazonka.IAM.Types.EntityDetails

Description

Documentation

data EntityDetails

An object that contains details about when the IAM entities (users or roles) were last used in an attempt to access the specified Amazon Web Services service.

This data type is a response element in the GetServiceLastAccessedDetailsWithEntities operation.

See: newEntityDetails smart constructor.

Constructors

EntityDetails'

Fields

lastAuthenticated :: Maybe ISO8601

The date and time, in ISO 8601 date-time format, when the authenticated entity last attempted to access Amazon Web Services. Amazon Web Services does not report unauthenticated requests.

This field is null if no IAM entities attempted to access the service within the reporting period.

entityInfo :: EntityInfo

The EntityInfo object that contains details about the entity (user or role).

newEntityDetails

Arguments

:: EntityInfo
$sel:entityInfo:EntityDetails'

-> EntityDetails

Create a value of EntityDetails with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:lastAuthenticated:EntityDetails', entityDetails_lastAuthenticated - The date and time, in ISO 8601 date-time format, when the authenticated entity last attempted to access Amazon Web Services. Amazon Web Services does not report unauthenticated requests.

This field is null if no IAM entities attempted to access the service within the reporting period.

$sel:entityInfo:EntityDetails', entityDetails_entityInfo - The EntityInfo object that contains details about the entity (user or role).

entityDetails_lastAuthenticated :: Lens' EntityDetails (Maybe UTCTime)

The date and time, in ISO 8601 date-time format, when the authenticated entity last attempted to access Amazon Web Services. Amazon Web Services does not report unauthenticated requests.

This field is null if no IAM entities attempted to access the service within the reporting period.

entityDetails_entityInfo :: Lens' EntityDetails EntityInfo

The EntityInfo object that contains details about the entity (user or role).