Amazonka.IAM.Types.Group

Description

Documentation

data Group

Contains information about an IAM group entity.

This data type is used as a response element in the following operations:

CreateGroup

GetGroup

ListGroups

See: newGroup smart constructor.

Constructors

Group'

Fields

path :: Text

The path to the group. For more information about paths, see IAM identifiers in the IAM User Guide.

groupName :: Text

The friendly name that identifies the group.

groupId :: Text

The stable and unique string identifying the group. For more information about IDs, see IAM identifiers in the IAM User Guide.

arn :: Text

The Amazon Resource Name (ARN) specifying the group. For more information about ARNs and how to use them in policies, see IAM identifiers in the IAM User Guide.

createDate :: ISO8601

The date and time, in ISO 8601 date-time format, when the group was created.

newGroup

Arguments

:: Text
$sel:path:Group'

-> Text
$sel:groupName:Group'

-> Text
$sel:groupId:Group'

-> Text
$sel:arn:Group'

-> UTCTime
$sel:createDate:Group'

-> Group

Create a value of Group with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:path:Group', group_path - The path to the group. For more information about paths, see IAM identifiers in the IAM User Guide.

$sel:groupName:Group', group_groupName - The friendly name that identifies the group.

$sel:groupId:Group', group_groupId - The stable and unique string identifying the group. For more information about IDs, see IAM identifiers in the IAM User Guide.

$sel:arn:Group', group_arn - The Amazon Resource Name (ARN) specifying the group. For more information about ARNs and how to use them in policies, see IAM identifiers in the IAM User Guide.

$sel:createDate:Group', group_createDate - The date and time, in ISO 8601 date-time format, when the group was created.

group_path :: Lens' Group Text

The path to the group. For more information about paths, see IAM identifiers in the IAM User Guide.

group_groupName :: Lens' Group Text

The friendly name that identifies the group.

group_groupId :: Lens' Group Text

The stable and unique string identifying the group. For more information about IDs, see IAM identifiers in the IAM User Guide.

group_arn :: Lens' Group Text

The Amazon Resource Name (ARN) specifying the group. For more information about ARNs and how to use them in policies, see IAM identifiers in the IAM User Guide.

group_createDate :: Lens' Group UTCTime

The date and time, in ISO 8601 date-time format, when the group was created.