Amazonka.IAM.Types.GroupDetail

Description

Documentation

data GroupDetail

Contains information about an IAM group, including all of the group's policies.

This data type is used as a response element in the GetAccountAuthorizationDetails operation.

See: newGroupDetail smart constructor.

Constructors

GroupDetail'

Fields

arn :: Maybe Text

path :: Maybe Text

The path to the group. For more information about paths, see IAM identifiers in the IAM User Guide.

createDate :: Maybe ISO8601

The date and time, in ISO 8601 date-time format, when the group was created.

groupId :: Maybe Text

The stable and unique string identifying the group. For more information about IDs, see IAM identifiers in the IAM User Guide.

groupPolicyList :: Maybe [PolicyDetail]

A list of the inline policies embedded in the group.

groupName :: Maybe Text

The friendly name that identifies the group.

attachedManagedPolicies :: Maybe [AttachedPolicy]

A list of the managed policies attached to the group.

newGroupDetail :: GroupDetail

Create a value of GroupDetail with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:arn:GroupDetail', groupDetail_arn - Undocumented member.

$sel:path:GroupDetail', groupDetail_path - The path to the group. For more information about paths, see IAM identifiers in the IAM User Guide.

$sel:createDate:GroupDetail', groupDetail_createDate - The date and time, in ISO 8601 date-time format, when the group was created.

$sel:groupId:GroupDetail', groupDetail_groupId - The stable and unique string identifying the group. For more information about IDs, see IAM identifiers in the IAM User Guide.

$sel:groupPolicyList:GroupDetail', groupDetail_groupPolicyList - A list of the inline policies embedded in the group.

$sel:groupName:GroupDetail', groupDetail_groupName - The friendly name that identifies the group.

$sel:attachedManagedPolicies:GroupDetail', groupDetail_attachedManagedPolicies - A list of the managed policies attached to the group.

groupDetail_arn :: Lens' GroupDetail (Maybe Text)

Undocumented member.

groupDetail_path :: Lens' GroupDetail (Maybe Text)

The path to the group. For more information about paths, see IAM identifiers in the IAM User Guide.

groupDetail_createDate :: Lens' GroupDetail (Maybe UTCTime)

The date and time, in ISO 8601 date-time format, when the group was created.

groupDetail_groupId :: Lens' GroupDetail (Maybe Text)

The stable and unique string identifying the group. For more information about IDs, see IAM identifiers in the IAM User Guide.

groupDetail_groupPolicyList :: Lens' GroupDetail (Maybe [PolicyDetail])

A list of the inline policies embedded in the group.

groupDetail_groupName :: Lens' GroupDetail (Maybe Text)

The friendly name that identifies the group.

groupDetail_attachedManagedPolicies :: Lens' GroupDetail (Maybe [AttachedPolicy])

A list of the managed policies attached to the group.