Amazonka.IAM.Types.InstanceProfile

Description

Documentation

data InstanceProfile

Contains information about an instance profile.

This data type is used as a response element in the following operations:

CreateInstanceProfile

GetInstanceProfile

ListInstanceProfiles

ListInstanceProfilesForRole

See: newInstanceProfile smart constructor.

Constructors

InstanceProfile'

Fields

tags :: Maybe [Tag]

A list of tags that are attached to the instance profile. For more information about tagging, see Tagging IAM resources in the IAM User Guide.

path :: Text

The path to the instance profile. For more information about paths, see IAM identifiers in the IAM User Guide.

instanceProfileName :: Text

The name identifying the instance profile.

instanceProfileId :: Text

The stable and unique string identifying the instance profile. For more information about IDs, see IAM identifiers in the IAM User Guide.

arn :: Text

The Amazon Resource Name (ARN) specifying the instance profile. For more information about ARNs and how to use them in policies, see IAM identifiers in the IAM User Guide.

createDate :: ISO8601

The date when the instance profile was created.

roles :: [Role]

The role associated with the instance profile.

newInstanceProfile

Arguments

:: Text
$sel:path:InstanceProfile'

-> Text
$sel:instanceProfileName:InstanceProfile'

-> Text
$sel:instanceProfileId:InstanceProfile'

-> Text
$sel:arn:InstanceProfile'

-> UTCTime
$sel:createDate:InstanceProfile'

-> InstanceProfile

Create a value of InstanceProfile with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:tags:InstanceProfile', instanceProfile_tags - A list of tags that are attached to the instance profile. For more information about tagging, see Tagging IAM resources in the IAM User Guide.

$sel:path:InstanceProfile', instanceProfile_path - The path to the instance profile. For more information about paths, see IAM identifiers in the IAM User Guide.

$sel:instanceProfileName:InstanceProfile', instanceProfile_instanceProfileName - The name identifying the instance profile.

$sel:instanceProfileId:InstanceProfile', instanceProfile_instanceProfileId - The stable and unique string identifying the instance profile. For more information about IDs, see IAM identifiers in the IAM User Guide.

$sel:arn:InstanceProfile', instanceProfile_arn - The Amazon Resource Name (ARN) specifying the instance profile. For more information about ARNs and how to use them in policies, see IAM identifiers in the IAM User Guide.

$sel:createDate:InstanceProfile', instanceProfile_createDate - The date when the instance profile was created.

$sel:roles:InstanceProfile', instanceProfile_roles - The role associated with the instance profile.

instanceProfile_tags :: Lens' InstanceProfile (Maybe [Tag])

A list of tags that are attached to the instance profile. For more information about tagging, see Tagging IAM resources in the IAM User Guide.

instanceProfile_path :: Lens' InstanceProfile Text

The path to the instance profile. For more information about paths, see IAM identifiers in the IAM User Guide.

instanceProfile_instanceProfileName :: Lens' InstanceProfile Text

The name identifying the instance profile.

instanceProfile_instanceProfileId :: Lens' InstanceProfile Text

The stable and unique string identifying the instance profile. For more information about IDs, see IAM identifiers in the IAM User Guide.

instanceProfile_arn :: Lens' InstanceProfile Text

The Amazon Resource Name (ARN) specifying the instance profile. For more information about ARNs and how to use them in policies, see IAM identifiers in the IAM User Guide.

instanceProfile_createDate :: Lens' InstanceProfile UTCTime

The date when the instance profile was created.

instanceProfile_roles :: Lens' InstanceProfile [Role]

The role associated with the instance profile.