Amazonka.IAM.Types.PolicyGrantingServiceAccess

Description

Documentation

data PolicyGrantingServiceAccess

Contains details about the permissions policies that are attached to the specified identity (user, group, or role).

This data type is an element of the ListPoliciesGrantingServiceAccessEntry object.

See: newPolicyGrantingServiceAccess smart constructor.

Constructors

PolicyGrantingServiceAccess'

Fields

entityName :: Maybe Text

The name of the entity (user or role) to which the inline policy is attached.

This field is null for managed policies. For more information about these policy types, see Managed policies and inline policies in the IAM User Guide.

entityType :: Maybe PolicyOwnerEntityType

The type of entity (user or role) that used the policy to access the service to which the inline policy is attached.

This field is null for managed policies. For more information about these policy types, see Managed policies and inline policies in the IAM User Guide.

policyArn :: Maybe Text

policyName :: Text

The policy name.

policyType :: PolicyType

The policy type. For more information about these policy types, see Managed policies and inline policies in the IAM User Guide.

newPolicyGrantingServiceAccess

Arguments

:: Text
$sel:policyName:PolicyGrantingServiceAccess'

-> PolicyType
$sel:policyType:PolicyGrantingServiceAccess'

-> PolicyGrantingServiceAccess

Create a value of PolicyGrantingServiceAccess with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:entityName:PolicyGrantingServiceAccess', policyGrantingServiceAccess_entityName - The name of the entity (user or role) to which the inline policy is attached.

This field is null for managed policies. For more information about these policy types, see Managed policies and inline policies in the IAM User Guide.

$sel:entityType:PolicyGrantingServiceAccess', policyGrantingServiceAccess_entityType - The type of entity (user or role) that used the policy to access the service to which the inline policy is attached.

This field is null for managed policies. For more information about these policy types, see Managed policies and inline policies in the IAM User Guide.

$sel:policyArn:PolicyGrantingServiceAccess', policyGrantingServiceAccess_policyArn - Undocumented member.

$sel:policyName:PolicyGrantingServiceAccess', policyGrantingServiceAccess_policyName - The policy name.

$sel:policyType:PolicyGrantingServiceAccess', policyGrantingServiceAccess_policyType - The policy type. For more information about these policy types, see Managed policies and inline policies in the IAM User Guide.

policyGrantingServiceAccess_entityName :: Lens' PolicyGrantingServiceAccess (Maybe Text)

The name of the entity (user or role) to which the inline policy is attached.

This field is null for managed policies. For more information about these policy types, see Managed policies and inline policies in the IAM User Guide.

policyGrantingServiceAccess_entityType :: Lens' PolicyGrantingServiceAccess (Maybe PolicyOwnerEntityType)

The type of entity (user or role) that used the policy to access the service to which the inline policy is attached.

This field is null for managed policies. For more information about these policy types, see Managed policies and inline policies in the IAM User Guide.

policyGrantingServiceAccess_policyArn :: Lens' PolicyGrantingServiceAccess (Maybe Text)

Undocumented member.

policyGrantingServiceAccess_policyName :: Lens' PolicyGrantingServiceAccess Text

The policy name.

policyGrantingServiceAccess_policyType :: Lens' PolicyGrantingServiceAccess PolicyType

The policy type. For more information about these policy types, see Managed policies and inline policies in the IAM User Guide.