Amazonka.IAM.Types.PolicyGroup

Description

Documentation

data PolicyGroup

Contains information about a group that a managed policy is attached to.

This data type is used as a response element in the ListEntitiesForPolicy operation.

For more information about managed policies, refer to Managed policies and inline policies in the IAM User Guide.

See: newPolicyGroup smart constructor.

Constructors

PolicyGroup'

Fields

groupId :: Maybe Text

The stable and unique string identifying the group. For more information about IDs, see IAM identifiers in the IAM User Guide.

groupName :: Maybe Text

The name (friendly name, not ARN) identifying the group.

newPolicyGroup :: PolicyGroup

Create a value of PolicyGroup with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:groupId:PolicyGroup', policyGroup_groupId - The stable and unique string identifying the group. For more information about IDs, see IAM identifiers in the IAM User Guide.

$sel:groupName:PolicyGroup', policyGroup_groupName - The name (friendly name, not ARN) identifying the group.

policyGroup_groupId :: Lens' PolicyGroup (Maybe Text)

The stable and unique string identifying the group. For more information about IDs, see IAM identifiers in the IAM User Guide.

policyGroup_groupName :: Lens' PolicyGroup (Maybe Text)

The name (friendly name, not ARN) identifying the group.