Amazonka.IAM.Types.PolicyUser

Description

Documentation

data PolicyUser

Contains information about a user that a managed policy is attached to.

This data type is used as a response element in the ListEntitiesForPolicy operation.

For more information about managed policies, refer to Managed policies and inline policies in the IAM User Guide.

See: newPolicyUser smart constructor.

Constructors

PolicyUser'

Fields

userName :: Maybe Text

The name (friendly name, not ARN) identifying the user.

userId :: Maybe Text

The stable and unique string identifying the user. For more information about IDs, see IAM identifiers in the IAM User Guide.

newPolicyUser :: PolicyUser

Create a value of PolicyUser with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:userName:PolicyUser', policyUser_userName - The name (friendly name, not ARN) identifying the user.

$sel:userId:PolicyUser', policyUser_userId - The stable and unique string identifying the user. For more information about IDs, see IAM identifiers in the IAM User Guide.

policyUser_userName :: Lens' PolicyUser (Maybe Text)

The name (friendly name, not ARN) identifying the user.

policyUser_userId :: Lens' PolicyUser (Maybe Text)

The stable and unique string identifying the user. For more information about IDs, see IAM identifiers in the IAM User Guide.