Amazonka.IAM.Types.PolicyVersion

Description

Documentation

data PolicyVersion

Contains information about a version of a managed policy.

This data type is used as a response element in the CreatePolicyVersion, GetPolicyVersion, ListPolicyVersions, and GetAccountAuthorizationDetails operations.

For more information about managed policies, refer to Managed policies and inline policies in the IAM User Guide.

See: newPolicyVersion smart constructor.

Constructors

PolicyVersion'

Fields

versionId :: Maybe Text

The identifier for the policy version.

Policy version identifiers always begin with v (always lowercase). When a policy is created, the first policy version is v1.

createDate :: Maybe ISO8601

The date and time, in ISO 8601 date-time format, when the policy version was created.

document :: Maybe Text

The policy document.

The policy document is returned in the response to the GetPolicyVersion and GetAccountAuthorizationDetails operations. It is not returned in the response to the CreatePolicyVersion or ListPolicyVersions operations.

The policy document returned in this structure is URL-encoded compliant with RFC 3986. You can use a URL decoding method to convert the policy back to plain JSON text. For example, if you use Java, you can use the decode method of the java.net.URLDecoder utility class in the Java SDK. Other languages and SDKs provide similar functionality.

isDefaultVersion :: Maybe Bool

Specifies whether the policy version is set as the policy's default version.

newPolicyVersion :: PolicyVersion

Create a value of PolicyVersion with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:versionId:PolicyVersion', policyVersion_versionId - The identifier for the policy version.

Policy version identifiers always begin with v (always lowercase). When a policy is created, the first policy version is v1.

$sel:createDate:PolicyVersion', policyVersion_createDate - The date and time, in ISO 8601 date-time format, when the policy version was created.

$sel:document:PolicyVersion', policyVersion_document - The policy document.

The policy document is returned in the response to the GetPolicyVersion and GetAccountAuthorizationDetails operations. It is not returned in the response to the CreatePolicyVersion or ListPolicyVersions operations.

The policy document returned in this structure is URL-encoded compliant with RFC 3986. You can use a URL decoding method to convert the policy back to plain JSON text. For example, if you use Java, you can use the decode method of the java.net.URLDecoder utility class in the Java SDK. Other languages and SDKs provide similar functionality.

$sel:isDefaultVersion:PolicyVersion', policyVersion_isDefaultVersion - Specifies whether the policy version is set as the policy's default version.

policyVersion_versionId :: Lens' PolicyVersion (Maybe Text)

The identifier for the policy version.

Policy version identifiers always begin with v (always lowercase). When a policy is created, the first policy version is v1.

policyVersion_createDate :: Lens' PolicyVersion (Maybe UTCTime)

The date and time, in ISO 8601 date-time format, when the policy version was created.

policyVersion_document :: Lens' PolicyVersion (Maybe Text)

The policy document.

The policy document is returned in the response to the GetPolicyVersion and GetAccountAuthorizationDetails operations. It is not returned in the response to the CreatePolicyVersion or ListPolicyVersions operations.

The policy document returned in this structure is URL-encoded compliant with RFC 3986. You can use a URL decoding method to convert the policy back to plain JSON text. For example, if you use Java, you can use the decode method of the java.net.URLDecoder utility class in the Java SDK. Other languages and SDKs provide similar functionality.

policyVersion_isDefaultVersion :: Lens' PolicyVersion (Maybe Bool)

Specifies whether the policy version is set as the policy's default version.