Amazonka.IAM.Types.SSHPublicKey

Description

Documentation

data SSHPublicKey

Contains information about an SSH public key.

This data type is used as a response element in the GetSSHPublicKey and UploadSSHPublicKey operations.

See: newSSHPublicKey smart constructor.

Constructors

SSHPublicKey'

Fields

uploadDate :: Maybe ISO8601

The date and time, in ISO 8601 date-time format, when the SSH public key was uploaded.

userName :: Text

The name of the IAM user associated with the SSH public key.

sSHPublicKeyId :: Text

The unique identifier for the SSH public key.

fingerprint :: Text

The MD5 message digest of the SSH public key.

sSHPublicKeyBody :: Text

The SSH public key.

status :: StatusType

The status of the SSH public key. Active means that the key can be used for authentication with an CodeCommit repository. Inactive means that the key cannot be used.

newSSHPublicKey

Arguments

:: Text
$sel:userName:SSHPublicKey'

-> Text
$sel:sSHPublicKeyId:SSHPublicKey'

-> Text
$sel:fingerprint:SSHPublicKey'

-> Text
$sel:sSHPublicKeyBody:SSHPublicKey'

-> StatusType
$sel:status:SSHPublicKey'

-> SSHPublicKey

Create a value of SSHPublicKey with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:uploadDate:SSHPublicKey', sSHPublicKey_uploadDate - The date and time, in ISO 8601 date-time format, when the SSH public key was uploaded.

$sel:userName:SSHPublicKey', sSHPublicKey_userName - The name of the IAM user associated with the SSH public key.

$sel:sSHPublicKeyId:SSHPublicKey', sSHPublicKey_sSHPublicKeyId - The unique identifier for the SSH public key.

$sel:fingerprint:SSHPublicKey', sSHPublicKey_fingerprint - The MD5 message digest of the SSH public key.

$sel:sSHPublicKeyBody:SSHPublicKey', sSHPublicKey_sSHPublicKeyBody - The SSH public key.

$sel:status:SSHPublicKey', sSHPublicKey_status - The status of the SSH public key. Active means that the key can be used for authentication with an CodeCommit repository. Inactive means that the key cannot be used.

sSHPublicKey_uploadDate :: Lens' SSHPublicKey (Maybe UTCTime)

The date and time, in ISO 8601 date-time format, when the SSH public key was uploaded.

sSHPublicKey_userName :: Lens' SSHPublicKey Text

The name of the IAM user associated with the SSH public key.

sSHPublicKey_sSHPublicKeyId :: Lens' SSHPublicKey Text

The unique identifier for the SSH public key.

sSHPublicKey_fingerprint :: Lens' SSHPublicKey Text

The MD5 message digest of the SSH public key.

sSHPublicKey_sSHPublicKeyBody :: Lens' SSHPublicKey Text

The SSH public key.

sSHPublicKey_status :: Lens' SSHPublicKey StatusType

The status of the SSH public key. Active means that the key can be used for authentication with an CodeCommit repository. Inactive means that the key cannot be used.