Amazonka.IAM.UpdateSAMLProvider

Description

Updates the metadata document for an existing SAML provider resource object.

This operation requires Signature Version 4.

Creating a Request

data UpdateSAMLProvider

See: newUpdateSAMLProvider smart constructor.

Constructors

UpdateSAMLProvider'

Fields

sAMLMetadataDocument :: Text

An XML document generated by an identity provider (IdP) that supports SAML 2.0. The document includes the issuer's name, expiration information, and keys that can be used to validate the SAML authentication response (assertions) that are received from the IdP. You must generate the metadata document using the identity management software that is used as your organization's IdP.

sAMLProviderArn :: Text

The Amazon Resource Name (ARN) of the SAML provider to update.

For more information about ARNs, see Amazon Resource Names (ARNs) in the Amazon Web Services General Reference.

newUpdateSAMLProvider

Arguments

:: Text
$sel:sAMLMetadataDocument:UpdateSAMLProvider'

-> Text
$sel:sAMLProviderArn:UpdateSAMLProvider'

-> UpdateSAMLProvider

Create a value of UpdateSAMLProvider with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:sAMLMetadataDocument:UpdateSAMLProvider', updateSAMLProvider_sAMLMetadataDocument - An XML document generated by an identity provider (IdP) that supports SAML 2.0. The document includes the issuer's name, expiration information, and keys that can be used to validate the SAML authentication response (assertions) that are received from the IdP. You must generate the metadata document using the identity management software that is used as your organization's IdP.

$sel:sAMLProviderArn:UpdateSAMLProvider', updateSAMLProvider_sAMLProviderArn - The Amazon Resource Name (ARN) of the SAML provider to update.

For more information about ARNs, see Amazon Resource Names (ARNs) in the Amazon Web Services General Reference.

Request Lenses

updateSAMLProvider_sAMLMetadataDocument :: Lens' UpdateSAMLProvider Text

An XML document generated by an identity provider (IdP) that supports SAML 2.0. The document includes the issuer's name, expiration information, and keys that can be used to validate the SAML authentication response (assertions) that are received from the IdP. You must generate the metadata document using the identity management software that is used as your organization's IdP.

updateSAMLProvider_sAMLProviderArn :: Lens' UpdateSAMLProvider Text

The Amazon Resource Name (ARN) of the SAML provider to update.

For more information about ARNs, see Amazon Resource Names (ARNs) in the Amazon Web Services General Reference.

Destructuring the Response

data UpdateSAMLProviderResponse

Contains the response to a successful UpdateSAMLProvider request.

See: newUpdateSAMLProviderResponse smart constructor.

Constructors

UpdateSAMLProviderResponse'

Fields

sAMLProviderArn :: Maybe Text

The Amazon Resource Name (ARN) of the SAML provider that was updated.

httpStatus :: Int

The response's http status code.

newUpdateSAMLProviderResponse

Arguments

:: Int
$sel:httpStatus:UpdateSAMLProviderResponse'

-> UpdateSAMLProviderResponse

Create a value of UpdateSAMLProviderResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:sAMLProviderArn:UpdateSAMLProvider', updateSAMLProviderResponse_sAMLProviderArn - The Amazon Resource Name (ARN) of the SAML provider that was updated.

$sel:httpStatus:UpdateSAMLProviderResponse', updateSAMLProviderResponse_httpStatus - The response's http status code.

Response Lenses

updateSAMLProviderResponse_sAMLProviderArn :: Lens' UpdateSAMLProviderResponse (Maybe Text)

The Amazon Resource Name (ARN) of the SAML provider that was updated.

updateSAMLProviderResponse_httpStatus :: Lens' UpdateSAMLProviderResponse Int

The response's http status code.