Amazonka.IAM.UpdateServiceSpecificCredential

Description

Sets the status of a service-specific credential to Active or Inactive. Service-specific credentials that are inactive cannot be used for authentication to the service. This operation can be used to disable a user's service-specific credential as part of a credential rotation work flow.

Creating a Request

data UpdateServiceSpecificCredential

See: newUpdateServiceSpecificCredential smart constructor.

Constructors

UpdateServiceSpecificCredential'

Fields

userName :: Maybe Text

The name of the IAM user associated with the service-specific credential. If you do not specify this value, then the operation assumes the user whose credentials are used to call the operation.

This parameter allows (through its regex pattern) a string of characters consisting of upper and lowercase alphanumeric characters with no spaces. You can also include any of the following characters: _+=,.@-

serviceSpecificCredentialId :: Text

The unique identifier of the service-specific credential.

This parameter allows (through its regex pattern) a string of characters that can consist of any upper or lowercased letter or digit.

status :: StatusType

The status to be assigned to the service-specific credential.

newUpdateServiceSpecificCredential

Arguments

:: Text
$sel:serviceSpecificCredentialId:UpdateServiceSpecificCredential'

-> StatusType
$sel:status:UpdateServiceSpecificCredential'

-> UpdateServiceSpecificCredential

Create a value of UpdateServiceSpecificCredential with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:userName:UpdateServiceSpecificCredential', updateServiceSpecificCredential_userName - The name of the IAM user associated with the service-specific credential. If you do not specify this value, then the operation assumes the user whose credentials are used to call the operation.

This parameter allows (through its regex pattern) a string of characters consisting of upper and lowercase alphanumeric characters with no spaces. You can also include any of the following characters: _+=,.@-

$sel:serviceSpecificCredentialId:UpdateServiceSpecificCredential', updateServiceSpecificCredential_serviceSpecificCredentialId - The unique identifier of the service-specific credential.

This parameter allows (through its regex pattern) a string of characters that can consist of any upper or lowercased letter or digit.

$sel:status:UpdateServiceSpecificCredential', updateServiceSpecificCredential_status - The status to be assigned to the service-specific credential.

Request Lenses

updateServiceSpecificCredential_userName :: Lens' UpdateServiceSpecificCredential (Maybe Text)

The name of the IAM user associated with the service-specific credential. If you do not specify this value, then the operation assumes the user whose credentials are used to call the operation.

This parameter allows (through its regex pattern) a string of characters consisting of upper and lowercase alphanumeric characters with no spaces. You can also include any of the following characters: _+=,.@-

updateServiceSpecificCredential_serviceSpecificCredentialId :: Lens' UpdateServiceSpecificCredential Text

The unique identifier of the service-specific credential.

This parameter allows (through its regex pattern) a string of characters that can consist of any upper or lowercased letter or digit.

updateServiceSpecificCredential_status :: Lens' UpdateServiceSpecificCredential StatusType

The status to be assigned to the service-specific credential.

Destructuring the Response

data UpdateServiceSpecificCredentialResponse

See: newUpdateServiceSpecificCredentialResponse smart constructor.

Constructors

UpdateServiceSpecificCredentialResponse'

newUpdateServiceSpecificCredentialResponse :: UpdateServiceSpecificCredentialResponse

Create a value of UpdateServiceSpecificCredentialResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.