Amazonka.IAM.UpdateSigningCertificate

Description

Changes the status of the specified user signing certificate from active to disabled, or vice versa. This operation can be used to disable an IAM user's signing certificate as part of a certificate rotation work flow.

If the UserName field is not specified, the user name is determined implicitly based on the Amazon Web Services access key ID used to sign the request. This operation works for access keys under the Amazon Web Services account. Consequently, you can use this operation to manage Amazon Web Services account root user credentials even if the Amazon Web Services account has no associated users.

Creating a Request

data UpdateSigningCertificate

See: newUpdateSigningCertificate smart constructor.

Constructors

UpdateSigningCertificate'

Fields

userName :: Maybe Text

The name of the IAM user the signing certificate belongs to.

This parameter allows (through its regex pattern) a string of characters consisting of upper and lowercase alphanumeric characters with no spaces. You can also include any of the following characters: _+=,.@-

certificateId :: Text

The ID of the signing certificate you want to update.

This parameter allows (through its regex pattern) a string of characters that can consist of any upper or lowercased letter or digit.

status :: StatusType

The status you want to assign to the certificate. Active means that the certificate can be used for programmatic calls to Amazon Web Services Inactive means that the certificate cannot be used.

newUpdateSigningCertificate

Arguments

:: Text
$sel:certificateId:UpdateSigningCertificate'

-> StatusType
$sel:status:UpdateSigningCertificate'

-> UpdateSigningCertificate

Create a value of UpdateSigningCertificate with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:userName:UpdateSigningCertificate', updateSigningCertificate_userName - The name of the IAM user the signing certificate belongs to.

This parameter allows (through its regex pattern) a string of characters consisting of upper and lowercase alphanumeric characters with no spaces. You can also include any of the following characters: _+=,.@-

$sel:certificateId:UpdateSigningCertificate', updateSigningCertificate_certificateId - The ID of the signing certificate you want to update.

This parameter allows (through its regex pattern) a string of characters that can consist of any upper or lowercased letter or digit.

$sel:status:UpdateSigningCertificate', updateSigningCertificate_status - The status you want to assign to the certificate. Active means that the certificate can be used for programmatic calls to Amazon Web Services Inactive means that the certificate cannot be used.

Request Lenses

updateSigningCertificate_userName :: Lens' UpdateSigningCertificate (Maybe Text)

The name of the IAM user the signing certificate belongs to.

This parameter allows (through its regex pattern) a string of characters consisting of upper and lowercase alphanumeric characters with no spaces. You can also include any of the following characters: _+=,.@-

updateSigningCertificate_certificateId :: Lens' UpdateSigningCertificate Text

The ID of the signing certificate you want to update.

This parameter allows (through its regex pattern) a string of characters that can consist of any upper or lowercased letter or digit.

updateSigningCertificate_status :: Lens' UpdateSigningCertificate StatusType

The status you want to assign to the certificate. Active means that the certificate can be used for programmatic calls to Amazon Web Services Inactive means that the certificate cannot be used.

Destructuring the Response

data UpdateSigningCertificateResponse

See: newUpdateSigningCertificateResponse smart constructor.

Constructors

UpdateSigningCertificateResponse'

newUpdateSigningCertificateResponse :: UpdateSigningCertificateResponse

Create a value of UpdateSigningCertificateResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.