Amazonka.IVS.CreateChannel

Description

Creates a new channel and an associated stream key to start streaming.

Creating a Request

data CreateChannel

See: newCreateChannel smart constructor.

Constructors

CreateChannel'

Fields

authorized :: Maybe Bool

Whether the channel is private (enabled for playback authorization). Default: false.

latencyMode :: Maybe ChannelLatencyMode

Channel latency mode. Use NORMAL to broadcast and deliver live video up to Full HD. Use LOW for near-real-time interaction with viewers. (Note: In the Amazon IVS console, LOW and NORMAL correspond to Ultra-low and Standard, respectively.) Default: LOW.

name :: Maybe Text

Channel name.

recordingConfigurationArn :: Maybe Text

Recording-configuration ARN. Default: "" (empty string, recording is disabled).

type' :: Maybe ChannelType

Channel type, which determines the allowable resolution and bitrate. /If you exceed the allowable resolution or bitrate, the stream probably will disconnect immediately./ Default: STANDARD. Valid values:

STANDARD: Multiple qualities are generated from the original input, to automatically give viewers the best experience for their devices and network conditions. Resolution can be up to 1080p and bitrate can be up to 8.5 Mbps. Audio is transcoded only for renditions 360p and below; above that, audio is passed through.

BASIC: Amazon IVS delivers the original input to viewers. The viewer’s video-quality choice is limited to the original input. Resolution can be up to 480p and bitrate can be up to 1.5 Mbps.

tags :: Maybe (HashMap Text Text)

Array of 1-50 maps, each of the form string:string (key:value).

newCreateChannel :: CreateChannel

Create a value of CreateChannel with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:authorized:CreateChannel', createChannel_authorized - Whether the channel is private (enabled for playback authorization). Default: false.

$sel:latencyMode:CreateChannel', createChannel_latencyMode - Channel latency mode. Use NORMAL to broadcast and deliver live video up to Full HD. Use LOW for near-real-time interaction with viewers. (Note: In the Amazon IVS console, LOW and NORMAL correspond to Ultra-low and Standard, respectively.) Default: LOW.

$sel:name:CreateChannel', createChannel_name - Channel name.

$sel:recordingConfigurationArn:CreateChannel', createChannel_recordingConfigurationArn - Recording-configuration ARN. Default: "" (empty string, recording is disabled).

$sel:type':CreateChannel', createChannel_type - Channel type, which determines the allowable resolution and bitrate. /If you exceed the allowable resolution or bitrate, the stream probably will disconnect immediately./ Default: STANDARD. Valid values:

STANDARD: Multiple qualities are generated from the original input, to automatically give viewers the best experience for their devices and network conditions. Resolution can be up to 1080p and bitrate can be up to 8.5 Mbps. Audio is transcoded only for renditions 360p and below; above that, audio is passed through.

BASIC: Amazon IVS delivers the original input to viewers. The viewer’s video-quality choice is limited to the original input. Resolution can be up to 480p and bitrate can be up to 1.5 Mbps.

$sel:tags:CreateChannel', createChannel_tags - Array of 1-50 maps, each of the form string:string (key:value).

Request Lenses

createChannel_authorized :: Lens' CreateChannel (Maybe Bool)

Whether the channel is private (enabled for playback authorization). Default: false.

createChannel_latencyMode :: Lens' CreateChannel (Maybe ChannelLatencyMode)

Channel latency mode. Use NORMAL to broadcast and deliver live video up to Full HD. Use LOW for near-real-time interaction with viewers. (Note: In the Amazon IVS console, LOW and NORMAL correspond to Ultra-low and Standard, respectively.) Default: LOW.

createChannel_name :: Lens' CreateChannel (Maybe Text)

Channel name.

createChannel_recordingConfigurationArn :: Lens' CreateChannel (Maybe Text)

Recording-configuration ARN. Default: "" (empty string, recording is disabled).

createChannel_type :: Lens' CreateChannel (Maybe ChannelType)

Channel type, which determines the allowable resolution and bitrate. /If you exceed the allowable resolution or bitrate, the stream probably will disconnect immediately./ Default: STANDARD. Valid values:

STANDARD: Multiple qualities are generated from the original input, to automatically give viewers the best experience for their devices and network conditions. Resolution can be up to 1080p and bitrate can be up to 8.5 Mbps. Audio is transcoded only for renditions 360p and below; above that, audio is passed through.

BASIC: Amazon IVS delivers the original input to viewers. The viewer’s video-quality choice is limited to the original input. Resolution can be up to 480p and bitrate can be up to 1.5 Mbps.

createChannel_tags :: Lens' CreateChannel (Maybe (HashMap Text Text))

Array of 1-50 maps, each of the form string:string (key:value).

Destructuring the Response

data CreateChannelResponse

See: newCreateChannelResponse smart constructor.

Constructors

CreateChannelResponse'

Fields

channel :: Maybe Channel

streamKey :: Maybe StreamKey

httpStatus :: Int

The response's http status code.

newCreateChannelResponse

Arguments

:: Int
$sel:httpStatus:CreateChannelResponse'

-> CreateChannelResponse

Create a value of CreateChannelResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:channel:CreateChannelResponse', createChannelResponse_channel -

$sel:streamKey:CreateChannelResponse', createChannelResponse_streamKey -

$sel:httpStatus:CreateChannelResponse', createChannelResponse_httpStatus - The response's http status code.

Response Lenses

createChannelResponse_channel :: Lens' CreateChannelResponse (Maybe Channel)

createChannelResponse_streamKey :: Lens' CreateChannelResponse (Maybe StreamKey)

createChannelResponse_httpStatus :: Lens' CreateChannelResponse Int

The response's http status code.