Amazonka.IVS.CreateRecordingConfiguration

Description

Creates a new recording configuration, used to enable recording to Amazon S3.

Known issue: In the us-east-1 region, if you use the Amazon Web Services CLI to create a recording configuration, it returns success even if the S3 bucket is in a different region. In this case, the state of the recording configuration is CREATE_FAILED (instead of ACTIVE). (In other regions, the CLI correctly returns failure if the bucket is in a different region.)

Workaround: Ensure that your S3 bucket is in the same region as the recording configuration. If you create a recording configuration in a different region as your S3 bucket, delete that recording configuration and create a new one with an S3 bucket from the correct region.

Creating a Request

data CreateRecordingConfiguration

See: newCreateRecordingConfiguration smart constructor.

Constructors

CreateRecordingConfiguration'

Fields

name :: Maybe Text

Recording-configuration name. The value does not need to be unique.

tags :: Maybe (HashMap Text Text)

Array of 1-50 maps, each of the form string:string (key:value).

destinationConfiguration :: DestinationConfiguration

A complex type that contains a destination configuration for where recorded video will be stored.

newCreateRecordingConfiguration

Arguments

:: DestinationConfiguration
$sel:destinationConfiguration:CreateRecordingConfiguration'

-> CreateRecordingConfiguration

Create a value of CreateRecordingConfiguration with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:name:CreateRecordingConfiguration', createRecordingConfiguration_name - Recording-configuration name. The value does not need to be unique.

$sel:tags:CreateRecordingConfiguration', createRecordingConfiguration_tags - Array of 1-50 maps, each of the form string:string (key:value).

$sel:destinationConfiguration:CreateRecordingConfiguration', createRecordingConfiguration_destinationConfiguration - A complex type that contains a destination configuration for where recorded video will be stored.

Request Lenses

createRecordingConfiguration_name :: Lens' CreateRecordingConfiguration (Maybe Text)

Recording-configuration name. The value does not need to be unique.

createRecordingConfiguration_tags :: Lens' CreateRecordingConfiguration (Maybe (HashMap Text Text))

Array of 1-50 maps, each of the form string:string (key:value).

createRecordingConfiguration_destinationConfiguration :: Lens' CreateRecordingConfiguration DestinationConfiguration

A complex type that contains a destination configuration for where recorded video will be stored.

Destructuring the Response

data CreateRecordingConfigurationResponse

See: newCreateRecordingConfigurationResponse smart constructor.

Constructors

CreateRecordingConfigurationResponse'

Fields

recordingConfiguration :: Maybe RecordingConfiguration

httpStatus :: Int

The response's http status code.

newCreateRecordingConfigurationResponse

Arguments

:: Int
$sel:httpStatus:CreateRecordingConfigurationResponse'

-> CreateRecordingConfigurationResponse

Create a value of CreateRecordingConfigurationResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:recordingConfiguration:CreateRecordingConfigurationResponse', createRecordingConfigurationResponse_recordingConfiguration -

$sel:httpStatus:CreateRecordingConfigurationResponse', createRecordingConfigurationResponse_httpStatus - The response's http status code.

Response Lenses

createRecordingConfigurationResponse_recordingConfiguration :: Lens' CreateRecordingConfigurationResponse (Maybe RecordingConfiguration)

createRecordingConfigurationResponse_httpStatus :: Lens' CreateRecordingConfigurationResponse Int

The response's http status code.