Amazonka.IVS.GetPlaybackKeyPair

Description

Gets a specified playback authorization key pair and returns the arn and fingerprint. The privateKey held by the caller can be used to generate viewer authorization tokens, to grant viewers access to private channels. For more information, see Setting Up Private Channels in the Amazon IVS User Guide.

Creating a Request

data GetPlaybackKeyPair

See: newGetPlaybackKeyPair smart constructor.

Constructors

GetPlaybackKeyPair'

Fields

arn :: Text

ARN of the key pair to be returned.

newGetPlaybackKeyPair

Arguments

:: Text
$sel:arn:GetPlaybackKeyPair'

-> GetPlaybackKeyPair

Create a value of GetPlaybackKeyPair with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:arn:GetPlaybackKeyPair', getPlaybackKeyPair_arn - ARN of the key pair to be returned.

Request Lenses

getPlaybackKeyPair_arn :: Lens' GetPlaybackKeyPair Text

ARN of the key pair to be returned.

Destructuring the Response

data GetPlaybackKeyPairResponse

See: newGetPlaybackKeyPairResponse smart constructor.

Constructors

GetPlaybackKeyPairResponse'

Fields

keyPair :: Maybe PlaybackKeyPair

httpStatus :: Int

The response's http status code.

newGetPlaybackKeyPairResponse

Arguments

:: Int
$sel:httpStatus:GetPlaybackKeyPairResponse'

-> GetPlaybackKeyPairResponse

Create a value of GetPlaybackKeyPairResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:keyPair:GetPlaybackKeyPairResponse', getPlaybackKeyPairResponse_keyPair -

$sel:httpStatus:GetPlaybackKeyPairResponse', getPlaybackKeyPairResponse_httpStatus - The response's http status code.

Response Lenses

getPlaybackKeyPairResponse_keyPair :: Lens' GetPlaybackKeyPairResponse (Maybe PlaybackKeyPair)

getPlaybackKeyPairResponse_httpStatus :: Lens' GetPlaybackKeyPairResponse Int

The response's http status code.