Amazonka.IVS.Lens

Description

Operations

PutMetadata

putMetadata_channelArn :: Lens' PutMetadata Text

ARN of the channel into which metadata is inserted. This channel must have an active stream.

putMetadata_metadata :: Lens' PutMetadata Text

Metadata to insert into the stream. Maximum: 1 KB per request.

ListRecordingConfigurations

listRecordingConfigurations_nextToken :: Lens' ListRecordingConfigurations (Maybe Text)

The first recording configuration to retrieve. This is used for pagination; see the nextToken response field.

listRecordingConfigurations_maxResults :: Lens' ListRecordingConfigurations (Maybe Natural)

Maximum number of recording configurations to return. Default: 50.

listRecordingConfigurationsResponse_nextToken :: Lens' ListRecordingConfigurationsResponse (Maybe Text)

If there are more recording configurations than maxResults, use nextToken in the request to get the next set.

listRecordingConfigurationsResponse_httpStatus :: Lens' ListRecordingConfigurationsResponse Int

The response's http status code.

listRecordingConfigurationsResponse_recordingConfigurations :: Lens' ListRecordingConfigurationsResponse [RecordingConfigurationSummary]

List of the matching recording configurations.

CreateStreamKey

createStreamKey_tags :: Lens' CreateStreamKey (Maybe (HashMap Text Text))

Array of 1-50 maps, each of the form string:string (key:value).

createStreamKey_channelArn :: Lens' CreateStreamKey Text

ARN of the channel for which to create the stream key.

createStreamKeyResponse_streamKey :: Lens' CreateStreamKeyResponse (Maybe StreamKey)

Stream key used to authenticate an RTMPS stream for ingestion.

createStreamKeyResponse_httpStatus :: Lens' CreateStreamKeyResponse Int

The response's http status code.

ImportPlaybackKeyPair

importPlaybackKeyPair_name :: Lens' ImportPlaybackKeyPair (Maybe Text)

Playback-key-pair name. The value does not need to be unique.

importPlaybackKeyPair_tags :: Lens' ImportPlaybackKeyPair (Maybe (HashMap Text Text))

Any tags provided with the request are added to the playback key pair tags.

importPlaybackKeyPair_publicKeyMaterial :: Lens' ImportPlaybackKeyPair Text

The public portion of a customer-generated key pair.

importPlaybackKeyPairResponse_keyPair :: Lens' ImportPlaybackKeyPairResponse (Maybe PlaybackKeyPair)

importPlaybackKeyPairResponse_httpStatus :: Lens' ImportPlaybackKeyPairResponse Int

The response's http status code.

ListChannels

listChannels_filterByName :: Lens' ListChannels (Maybe Text)

Filters the channel list to match the specified name.

listChannels_nextToken :: Lens' ListChannels (Maybe Text)

The first channel to retrieve. This is used for pagination; see the nextToken response field.

listChannels_filterByRecordingConfigurationArn :: Lens' ListChannels (Maybe Text)

Filters the channel list to match the specified recording-configuration ARN.

listChannels_maxResults :: Lens' ListChannels (Maybe Natural)

Maximum number of channels to return. Default: 50.

listChannelsResponse_nextToken :: Lens' ListChannelsResponse (Maybe Text)

If there are more channels than maxResults, use nextToken in the request to get the next set.

listChannelsResponse_httpStatus :: Lens' ListChannelsResponse Int

The response's http status code.

listChannelsResponse_channels :: Lens' ListChannelsResponse [ChannelSummary]

List of the matching channels.

ListTagsForResource

listTagsForResource_resourceArn :: Lens' ListTagsForResource Text

The ARN of the resource to be retrieved.

listTagsForResourceResponse_httpStatus :: Lens' ListTagsForResourceResponse Int

The response's http status code.

listTagsForResourceResponse_tags :: Lens' ListTagsForResourceResponse (HashMap Text Text)

DeleteChannel

deleteChannel_arn :: Lens' DeleteChannel Text

ARN of the channel to be deleted.

UpdateChannel

updateChannel_authorized :: Lens' UpdateChannel (Maybe Bool)

Whether the channel is private (enabled for playback authorization).

updateChannel_latencyMode :: Lens' UpdateChannel (Maybe ChannelLatencyMode)

Channel latency mode. Use NORMAL to broadcast and deliver live video up to Full HD. Use LOW for near-real-time interaction with viewers. (Note: In the Amazon IVS console, LOW and NORMAL correspond to Ultra-low and Standard, respectively.)

updateChannel_name :: Lens' UpdateChannel (Maybe Text)

Channel name.

updateChannel_recordingConfigurationArn :: Lens' UpdateChannel (Maybe Text)

Recording-configuration ARN. If this is set to an empty string, recording is disabled. A value other than an empty string indicates that recording is enabled

updateChannel_type :: Lens' UpdateChannel (Maybe ChannelType)

Channel type, which determines the allowable resolution and bitrate. /If you exceed the allowable resolution or bitrate, the stream probably will disconnect immediately/. Valid values:

STANDARD: Multiple qualities are generated from the original input, to automatically give viewers the best experience for their devices and network conditions. Resolution can be up to 1080p and bitrate can be up to 8.5 Mbps. Audio is transcoded only for renditions 360p and below; above that, audio is passed through.

BASIC: Amazon IVS delivers the original input to viewers. The viewer’s video-quality choice is limited to the original input. Resolution can be up to 480p and bitrate can be up to 1.5 Mbps.

updateChannel_arn :: Lens' UpdateChannel Text

ARN of the channel to be updated.

updateChannelResponse_channel :: Lens' UpdateChannelResponse (Maybe Channel)

Undocumented member.

updateChannelResponse_httpStatus :: Lens' UpdateChannelResponse Int

The response's http status code.

GetStreamKey

getStreamKey_arn :: Lens' GetStreamKey Text

ARN for the stream key to be retrieved.

getStreamKeyResponse_streamKey :: Lens' GetStreamKeyResponse (Maybe StreamKey)

getStreamKeyResponse_httpStatus :: Lens' GetStreamKeyResponse Int

The response's http status code.

DeletePlaybackKeyPair

deletePlaybackKeyPair_arn :: Lens' DeletePlaybackKeyPair Text

ARN of the key pair to be deleted.

deletePlaybackKeyPairResponse_httpStatus :: Lens' DeletePlaybackKeyPairResponse Int

The response's http status code.

BatchGetStreamKey

batchGetStreamKey_arns :: Lens' BatchGetStreamKey (NonEmpty Text)

Array of ARNs, one per channel.

batchGetStreamKeyResponse_streamKeys :: Lens' BatchGetStreamKeyResponse (Maybe [StreamKey])

batchGetStreamKeyResponse_errors :: Lens' BatchGetStreamKeyResponse (Maybe [BatchError])

batchGetStreamKeyResponse_httpStatus :: Lens' BatchGetStreamKeyResponse Int

The response's http status code.

GetPlaybackKeyPair

getPlaybackKeyPair_arn :: Lens' GetPlaybackKeyPair Text

ARN of the key pair to be returned.

getPlaybackKeyPairResponse_keyPair :: Lens' GetPlaybackKeyPairResponse (Maybe PlaybackKeyPair)

getPlaybackKeyPairResponse_httpStatus :: Lens' GetPlaybackKeyPairResponse Int

The response's http status code.

DeleteRecordingConfiguration

deleteRecordingConfiguration_arn :: Lens' DeleteRecordingConfiguration Text

ARN of the recording configuration to be deleted.

StopStream

stopStream_channelArn :: Lens' StopStream Text

ARN of the channel for which the stream is to be stopped.

stopStreamResponse_httpStatus :: Lens' StopStreamResponse Int

The response's http status code.

CreateChannel

createChannel_authorized :: Lens' CreateChannel (Maybe Bool)

Whether the channel is private (enabled for playback authorization). Default: false.

createChannel_latencyMode :: Lens' CreateChannel (Maybe ChannelLatencyMode)

Channel latency mode. Use NORMAL to broadcast and deliver live video up to Full HD. Use LOW for near-real-time interaction with viewers. (Note: In the Amazon IVS console, LOW and NORMAL correspond to Ultra-low and Standard, respectively.) Default: LOW.

createChannel_name :: Lens' CreateChannel (Maybe Text)

Channel name.

createChannel_recordingConfigurationArn :: Lens' CreateChannel (Maybe Text)

Recording-configuration ARN. Default: "" (empty string, recording is disabled).

createChannel_type :: Lens' CreateChannel (Maybe ChannelType)

Channel type, which determines the allowable resolution and bitrate. /If you exceed the allowable resolution or bitrate, the stream probably will disconnect immediately./ Default: STANDARD. Valid values:

STANDARD: Multiple qualities are generated from the original input, to automatically give viewers the best experience for their devices and network conditions. Resolution can be up to 1080p and bitrate can be up to 8.5 Mbps. Audio is transcoded only for renditions 360p and below; above that, audio is passed through.

BASIC: Amazon IVS delivers the original input to viewers. The viewer’s video-quality choice is limited to the original input. Resolution can be up to 480p and bitrate can be up to 1.5 Mbps.

createChannel_tags :: Lens' CreateChannel (Maybe (HashMap Text Text))

Array of 1-50 maps, each of the form string:string (key:value).

createChannelResponse_channel :: Lens' CreateChannelResponse (Maybe Channel)

createChannelResponse_streamKey :: Lens' CreateChannelResponse (Maybe StreamKey)

createChannelResponse_httpStatus :: Lens' CreateChannelResponse Int

The response's http status code.

DeleteStreamKey

deleteStreamKey_arn :: Lens' DeleteStreamKey Text

ARN of the stream key to be deleted.

GetStream

getStream_channelArn :: Lens' GetStream Text

Channel ARN for stream to be accessed.

getStreamResponse_stream :: Lens' GetStreamResponse (Maybe Stream)

getStreamResponse_httpStatus :: Lens' GetStreamResponse Int

The response's http status code.

ListStreamKeys

listStreamKeys_nextToken :: Lens' ListStreamKeys (Maybe Text)

The first stream key to retrieve. This is used for pagination; see the nextToken response field.

listStreamKeys_maxResults :: Lens' ListStreamKeys (Maybe Natural)

Maximum number of streamKeys to return. Default: 50.

listStreamKeys_channelArn :: Lens' ListStreamKeys Text

Channel ARN used to filter the list.

listStreamKeysResponse_nextToken :: Lens' ListStreamKeysResponse (Maybe Text)

If there are more stream keys than maxResults, use nextToken in the request to get the next set.

listStreamKeysResponse_httpStatus :: Lens' ListStreamKeysResponse Int

The response's http status code.

listStreamKeysResponse_streamKeys :: Lens' ListStreamKeysResponse [StreamKeySummary]

List of stream keys.

GetChannel

getChannel_arn :: Lens' GetChannel Text

ARN of the channel for which the configuration is to be retrieved.

getChannelResponse_channel :: Lens' GetChannelResponse (Maybe Channel)

getChannelResponse_httpStatus :: Lens' GetChannelResponse Int

The response's http status code.

ListStreams

listStreams_nextToken :: Lens' ListStreams (Maybe Text)

The first stream to retrieve. This is used for pagination; see the nextToken response field.

listStreams_maxResults :: Lens' ListStreams (Maybe Natural)

Maximum number of streams to return. Default: 50.

listStreamsResponse_nextToken :: Lens' ListStreamsResponse (Maybe Text)

If there are more streams than maxResults, use nextToken in the request to get the next set.

listStreamsResponse_httpStatus :: Lens' ListStreamsResponse Int

The response's http status code.

listStreamsResponse_streams :: Lens' ListStreamsResponse [StreamSummary]

List of streams.

TagResource

tagResource_resourceArn :: Lens' TagResource Text

ARN of the resource for which tags are to be added or updated.

tagResource_tags :: Lens' TagResource (HashMap Text Text)

Array of tags to be added or updated.

tagResourceResponse_httpStatus :: Lens' TagResourceResponse Int

The response's http status code.

GetRecordingConfiguration

getRecordingConfiguration_arn :: Lens' GetRecordingConfiguration Text

ARN of the recording configuration to be retrieved.

getRecordingConfigurationResponse_recordingConfiguration :: Lens' GetRecordingConfigurationResponse (Maybe RecordingConfiguration)

getRecordingConfigurationResponse_httpStatus :: Lens' GetRecordingConfigurationResponse Int

The response's http status code.

ListPlaybackKeyPairs

listPlaybackKeyPairs_nextToken :: Lens' ListPlaybackKeyPairs (Maybe Text)

Maximum number of key pairs to return.

listPlaybackKeyPairs_maxResults :: Lens' ListPlaybackKeyPairs (Maybe Natural)

The first key pair to retrieve. This is used for pagination; see the nextToken response field. Default: 50.

listPlaybackKeyPairsResponse_nextToken :: Lens' ListPlaybackKeyPairsResponse (Maybe Text)

If there are more key pairs than maxResults, use nextToken in the request to get the next set.

listPlaybackKeyPairsResponse_httpStatus :: Lens' ListPlaybackKeyPairsResponse Int

The response's http status code.

listPlaybackKeyPairsResponse_keyPairs :: Lens' ListPlaybackKeyPairsResponse [PlaybackKeyPairSummary]

List of key pairs.

UntagResource

untagResource_resourceArn :: Lens' UntagResource Text

ARN of the resource for which tags are to be removed.

untagResource_tagKeys :: Lens' UntagResource [Text]

Array of tags to be removed.

untagResourceResponse_httpStatus :: Lens' UntagResourceResponse Int

The response's http status code.

CreateRecordingConfiguration

createRecordingConfiguration_name :: Lens' CreateRecordingConfiguration (Maybe Text)

Recording-configuration name. The value does not need to be unique.

createRecordingConfiguration_tags :: Lens' CreateRecordingConfiguration (Maybe (HashMap Text Text))

Array of 1-50 maps, each of the form string:string (key:value).

createRecordingConfiguration_destinationConfiguration :: Lens' CreateRecordingConfiguration DestinationConfiguration

A complex type that contains a destination configuration for where recorded video will be stored.

createRecordingConfigurationResponse_recordingConfiguration :: Lens' CreateRecordingConfigurationResponse (Maybe RecordingConfiguration)

createRecordingConfigurationResponse_httpStatus :: Lens' CreateRecordingConfigurationResponse Int

The response's http status code.

BatchGetChannel

batchGetChannel_arns :: Lens' BatchGetChannel (NonEmpty Text)

Array of ARNs, one per channel.

batchGetChannelResponse_channels :: Lens' BatchGetChannelResponse (Maybe [Channel])

batchGetChannelResponse_errors :: Lens' BatchGetChannelResponse (Maybe [BatchError])

Each error object is related to a specific ARN in the request.

batchGetChannelResponse_httpStatus :: Lens' BatchGetChannelResponse Int

The response's http status code.

Types

BatchError

batchError_arn :: Lens' BatchError (Maybe Text)

Channel ARN.

batchError_code :: Lens' BatchError (Maybe Text)

Error code.

batchError_message :: Lens' BatchError (Maybe Text)

Error message, determined by the application.

Channel

channel_playbackUrl :: Lens' Channel (Maybe Text)

Channel playback URL.

channel_authorized :: Lens' Channel (Maybe Bool)

Whether the channel is private (enabled for playback authorization). Default: false.

channel_arn :: Lens' Channel (Maybe Text)

Channel ARN.

channel_latencyMode :: Lens' Channel (Maybe ChannelLatencyMode)

Channel latency mode. Use NORMAL to broadcast and deliver live video up to Full HD. Use LOW for near-real-time interaction with viewers. Default: LOW. (Note: In the Amazon IVS console, LOW and NORMAL correspond to Ultra-low and Standard, respectively.)

channel_name :: Lens' Channel (Maybe Text)

Channel name.

channel_recordingConfigurationArn :: Lens' Channel (Maybe Text)

Recording-configuration ARN. A value other than an empty string indicates that recording is enabled. Default: "" (empty string, recording is disabled).

channel_type :: Lens' Channel (Maybe ChannelType)

Channel type, which determines the allowable resolution and bitrate. /If you exceed the allowable resolution or bitrate, the stream probably will disconnect immediately./ Default: STANDARD. Valid values:

STANDARD: Multiple qualities are generated from the original input, to automatically give viewers the best experience for their devices and network conditions. Resolution can be up to 1080p and bitrate can be up to 8.5 Mbps. Audio is transcoded only for renditions 360p and below; above that, audio is passed through.

BASIC: Amazon IVS delivers the original input to viewers. The viewer’s video-quality choice is limited to the original input. Resolution can be up to 480p and bitrate can be up to 1.5 Mbps.

channel_tags :: Lens' Channel (Maybe (HashMap Text Text))

Array of 1-50 maps, each of the form string:string (key:value).

channel_ingestEndpoint :: Lens' Channel (Maybe Text)

Channel ingest endpoint, part of the definition of an ingest server, used when you set up streaming software.

ChannelSummary

channelSummary_authorized :: Lens' ChannelSummary (Maybe Bool)

Whether the channel is private (enabled for playback authorization). Default: false.

channelSummary_arn :: Lens' ChannelSummary (Maybe Text)

Channel ARN.

channelSummary_latencyMode :: Lens' ChannelSummary (Maybe ChannelLatencyMode)

Channel latency mode. Use NORMAL to broadcast and deliver live video up to Full HD. Use LOW for near-real-time interaction with viewers. Default: LOW. (Note: In the Amazon IVS console, LOW and NORMAL correspond to Ultra-low and Standard, respectively.)

channelSummary_name :: Lens' ChannelSummary (Maybe Text)

Channel name.

channelSummary_recordingConfigurationArn :: Lens' ChannelSummary (Maybe Text)

Recording-configuration ARN. A value other than an empty string indicates that recording is enabled. Default: "" (empty string, recording is disabled).

channelSummary_tags :: Lens' ChannelSummary (Maybe (HashMap Text Text))

Array of 1-50 maps, each of the form string:string (key:value).

DestinationConfiguration

destinationConfiguration_s3 :: Lens' DestinationConfiguration (Maybe S3DestinationConfiguration)

An S3 destination configuration where recorded videos will be stored.

PlaybackKeyPair

playbackKeyPair_arn :: Lens' PlaybackKeyPair (Maybe Text)

Key-pair ARN.

playbackKeyPair_fingerprint :: Lens' PlaybackKeyPair (Maybe Text)

Key-pair identifier.

playbackKeyPair_name :: Lens' PlaybackKeyPair (Maybe Text)

Playback-key-pair name. The value does not need to be unique.

playbackKeyPair_tags :: Lens' PlaybackKeyPair (Maybe (HashMap Text Text))

Array of 1-50 maps, each of the form string:string (key:value).

PlaybackKeyPairSummary

playbackKeyPairSummary_arn :: Lens' PlaybackKeyPairSummary (Maybe Text)

Key-pair ARN.

playbackKeyPairSummary_name :: Lens' PlaybackKeyPairSummary (Maybe Text)

Playback-key-pair name. The value does not need to be unique.

playbackKeyPairSummary_tags :: Lens' PlaybackKeyPairSummary (Maybe (HashMap Text Text))

Array of 1-50 maps, each of the form string:string (key:value).

RecordingConfiguration

recordingConfiguration_name :: Lens' RecordingConfiguration (Maybe Text)

Recording-configuration name. The value does not need to be unique.

recordingConfiguration_tags :: Lens' RecordingConfiguration (Maybe (HashMap Text Text))

Array of 1-50 maps, each of the form string:string (key:value).

recordingConfiguration_arn :: Lens' RecordingConfiguration Text

Recording-configuration ARN.

recordingConfiguration_destinationConfiguration :: Lens' RecordingConfiguration DestinationConfiguration

A complex type that contains information about where recorded video will be stored.

recordingConfiguration_state :: Lens' RecordingConfiguration RecordingConfigurationState

Indicates the current state of the recording configuration. When the state is ACTIVE, the configuration is ready for recording a channel stream.

RecordingConfigurationSummary

recordingConfigurationSummary_name :: Lens' RecordingConfigurationSummary (Maybe Text)

Recording-configuration name. The value does not need to be unique.

recordingConfigurationSummary_tags :: Lens' RecordingConfigurationSummary (Maybe (HashMap Text Text))

Array of 1-50 maps, each of the form string:string (key:value).

recordingConfigurationSummary_arn :: Lens' RecordingConfigurationSummary Text

Recording-configuration ARN.

recordingConfigurationSummary_destinationConfiguration :: Lens' RecordingConfigurationSummary DestinationConfiguration

A complex type that contains information about where recorded video will be stored.

recordingConfigurationSummary_state :: Lens' RecordingConfigurationSummary RecordingConfigurationState

Indicates the current state of the recording configuration. When the state is ACTIVE, the configuration is ready for recording a channel stream.

S3DestinationConfiguration

s3DestinationConfiguration_bucketName :: Lens' S3DestinationConfiguration Text

Location (S3 bucket name) where recorded videos will be stored.

Stream

stream_playbackUrl :: Lens' Stream (Maybe Text)

URL of the master playlist, required by the video player to play the HLS stream.

stream_state :: Lens' Stream (Maybe StreamState)

The stream’s state.

stream_startTime :: Lens' Stream (Maybe UTCTime)

Time of the stream’s start. This is an ISO 8601 timestamp returned as a string.

stream_channelArn :: Lens' Stream (Maybe Text)

Channel ARN for the stream.

stream_viewerCount :: Lens' Stream (Maybe Integer)

A count of concurrent views of the stream. Typically, a new view appears in viewerCount within 15 seconds of when video playback starts and a view is removed from viewerCount within 1 minute of when video playback ends. A value of -1 indicates that the request timed out; in this case, retry.

stream_health :: Lens' Stream (Maybe StreamHealth)

The stream’s health.

StreamKey

streamKey_arn :: Lens' StreamKey (Maybe Text)

Stream-key ARN.

streamKey_value :: Lens' StreamKey (Maybe Text)

Stream-key value.

streamKey_channelArn :: Lens' StreamKey (Maybe Text)

Channel ARN for the stream.

streamKey_tags :: Lens' StreamKey (Maybe (HashMap Text Text))

Array of 1-50 maps, each of the form string:string (key:value).

StreamKeySummary

streamKeySummary_arn :: Lens' StreamKeySummary (Maybe Text)

Stream-key ARN.

streamKeySummary_channelArn :: Lens' StreamKeySummary (Maybe Text)

Channel ARN for the stream.

streamKeySummary_tags :: Lens' StreamKeySummary (Maybe (HashMap Text Text))

Array of 1-50 maps, each of the form string:string (key:value).

StreamSummary

streamSummary_state :: Lens' StreamSummary (Maybe StreamState)

The stream’s state.

streamSummary_startTime :: Lens' StreamSummary (Maybe UTCTime)

Time of the stream’s start. This is an ISO 8601 timestamp returned as a string.

streamSummary_channelArn :: Lens' StreamSummary (Maybe Text)

Channel ARN for the stream.

streamSummary_viewerCount :: Lens' StreamSummary (Maybe Integer)

A count of concurrent views of the stream. Typically, a new view appears in viewerCount within 15 seconds of when video playback starts and a view is removed from viewerCount within 1 minute of when video playback ends. A value of -1 indicates that the request timed out; in this case, retry.

streamSummary_health :: Lens' StreamSummary (Maybe StreamHealth)

The stream’s health.