Amazonka.IVS.Types

Description

Service Configuration

defaultService :: Service

API version 2020-07-14 of the Amazon Interactive Video Service SDK configuration.

Errors

_ValidationException :: AsError a => Getting (First ServiceError) a ServiceError

_AccessDeniedException :: AsError a => Getting (First ServiceError) a ServiceError

_StreamUnavailable :: AsError a => Getting (First ServiceError) a ServiceError

_PendingVerification :: AsError a => Getting (First ServiceError) a ServiceError

_ChannelNotBroadcasting :: AsError a => Getting (First ServiceError) a ServiceError

_ConflictException :: AsError a => Getting (First ServiceError) a ServiceError

_ServiceQuotaExceededException :: AsError a => Getting (First ServiceError) a ServiceError

_ThrottlingException :: AsError a => Getting (First ServiceError) a ServiceError

_InternalServerException :: AsError a => Getting (First ServiceError) a ServiceError

_ResourceNotFoundException :: AsError a => Getting (First ServiceError) a ServiceError

ChannelLatencyMode

newtype ChannelLatencyMode

Constructors

ChannelLatencyMode'

Fields

fromChannelLatencyMode :: Text

Bundled Patterns

pattern ChannelLatencyMode_LOW :: ChannelLatencyMode

pattern ChannelLatencyMode_NORMAL :: ChannelLatencyMode

ChannelType

newtype ChannelType

Constructors

ChannelType'

Fields

fromChannelType :: Text

Bundled Patterns

pattern ChannelType_BASIC :: ChannelType

pattern ChannelType_STANDARD :: ChannelType

RecordingConfigurationState

newtype RecordingConfigurationState

Constructors

RecordingConfigurationState'

Fields

fromRecordingConfigurationState :: Text

Bundled Patterns

pattern RecordingConfigurationState_ACTIVE :: RecordingConfigurationState

pattern RecordingConfigurationState_CREATE_FAILED :: RecordingConfigurationState

pattern RecordingConfigurationState_CREATING :: RecordingConfigurationState

StreamHealth

newtype StreamHealth

Constructors

StreamHealth'

Fields

fromStreamHealth :: Text

Bundled Patterns

pattern StreamHealth_HEALTHY :: StreamHealth

pattern StreamHealth_STARVING :: StreamHealth

pattern StreamHealth_UNKNOWN :: StreamHealth

StreamState

newtype StreamState

Constructors

StreamState'

Fields

fromStreamState :: Text

Bundled Patterns

pattern StreamState_LIVE :: StreamState

pattern StreamState_OFFLINE :: StreamState

BatchError

data BatchError

Error related to a specific channel, specified by its ARN.

See: newBatchError smart constructor.

Constructors

BatchError'

Fields

arn :: Maybe Text

Channel ARN.

code :: Maybe Text

Error code.

message :: Maybe Text

Error message, determined by the application.

newBatchError :: BatchError

Create a value of BatchError with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:arn:BatchError', batchError_arn - Channel ARN.

$sel:code:BatchError', batchError_code - Error code.

$sel:message:BatchError', batchError_message - Error message, determined by the application.

batchError_arn :: Lens' BatchError (Maybe Text)

Channel ARN.

batchError_code :: Lens' BatchError (Maybe Text)

Error code.

batchError_message :: Lens' BatchError (Maybe Text)

Error message, determined by the application.

Channel

data Channel

Object specifying a channel.

See: newChannel smart constructor.

Constructors

Channel'

Fields

playbackUrl :: Maybe Text

Channel playback URL.

authorized :: Maybe Bool

Whether the channel is private (enabled for playback authorization). Default: false.

arn :: Maybe Text

Channel ARN.

latencyMode :: Maybe ChannelLatencyMode

Channel latency mode. Use NORMAL to broadcast and deliver live video up to Full HD. Use LOW for near-real-time interaction with viewers. Default: LOW. (Note: In the Amazon IVS console, LOW and NORMAL correspond to Ultra-low and Standard, respectively.)

name :: Maybe Text

Channel name.

recordingConfigurationArn :: Maybe Text

Recording-configuration ARN. A value other than an empty string indicates that recording is enabled. Default: "" (empty string, recording is disabled).

type' :: Maybe ChannelType

Channel type, which determines the allowable resolution and bitrate. /If you exceed the allowable resolution or bitrate, the stream probably will disconnect immediately./ Default: STANDARD. Valid values:

STANDARD: Multiple qualities are generated from the original input, to automatically give viewers the best experience for their devices and network conditions. Resolution can be up to 1080p and bitrate can be up to 8.5 Mbps. Audio is transcoded only for renditions 360p and below; above that, audio is passed through.

BASIC: Amazon IVS delivers the original input to viewers. The viewer’s video-quality choice is limited to the original input. Resolution can be up to 480p and bitrate can be up to 1.5 Mbps.

tags :: Maybe (HashMap Text Text)

Array of 1-50 maps, each of the form string:string (key:value).

ingestEndpoint :: Maybe Text

Channel ingest endpoint, part of the definition of an ingest server, used when you set up streaming software.

newChannel :: Channel

Create a value of Channel with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:playbackUrl:Channel', channel_playbackUrl - Channel playback URL.

$sel:authorized:Channel', channel_authorized - Whether the channel is private (enabled for playback authorization). Default: false.

$sel:arn:Channel', channel_arn - Channel ARN.

$sel:latencyMode:Channel', channel_latencyMode - Channel latency mode. Use NORMAL to broadcast and deliver live video up to Full HD. Use LOW for near-real-time interaction with viewers. Default: LOW. (Note: In the Amazon IVS console, LOW and NORMAL correspond to Ultra-low and Standard, respectively.)

$sel:name:Channel', channel_name - Channel name.

$sel:recordingConfigurationArn:Channel', channel_recordingConfigurationArn - Recording-configuration ARN. A value other than an empty string indicates that recording is enabled. Default: "" (empty string, recording is disabled).

$sel:type':Channel', channel_type - Channel type, which determines the allowable resolution and bitrate. /If you exceed the allowable resolution or bitrate, the stream probably will disconnect immediately./ Default: STANDARD. Valid values:

STANDARD: Multiple qualities are generated from the original input, to automatically give viewers the best experience for their devices and network conditions. Resolution can be up to 1080p and bitrate can be up to 8.5 Mbps. Audio is transcoded only for renditions 360p and below; above that, audio is passed through.

BASIC: Amazon IVS delivers the original input to viewers. The viewer’s video-quality choice is limited to the original input. Resolution can be up to 480p and bitrate can be up to 1.5 Mbps.

$sel:tags:Channel', channel_tags - Array of 1-50 maps, each of the form string:string (key:value).

$sel:ingestEndpoint:Channel', channel_ingestEndpoint - Channel ingest endpoint, part of the definition of an ingest server, used when you set up streaming software.

channel_playbackUrl :: Lens' Channel (Maybe Text)

Channel playback URL.

channel_authorized :: Lens' Channel (Maybe Bool)

Whether the channel is private (enabled for playback authorization). Default: false.

channel_arn :: Lens' Channel (Maybe Text)

Channel ARN.

channel_latencyMode :: Lens' Channel (Maybe ChannelLatencyMode)

Channel latency mode. Use NORMAL to broadcast and deliver live video up to Full HD. Use LOW for near-real-time interaction with viewers. Default: LOW. (Note: In the Amazon IVS console, LOW and NORMAL correspond to Ultra-low and Standard, respectively.)

channel_name :: Lens' Channel (Maybe Text)

Channel name.

channel_recordingConfigurationArn :: Lens' Channel (Maybe Text)

Recording-configuration ARN. A value other than an empty string indicates that recording is enabled. Default: "" (empty string, recording is disabled).

channel_type :: Lens' Channel (Maybe ChannelType)

Channel type, which determines the allowable resolution and bitrate. /If you exceed the allowable resolution or bitrate, the stream probably will disconnect immediately./ Default: STANDARD. Valid values:

STANDARD: Multiple qualities are generated from the original input, to automatically give viewers the best experience for their devices and network conditions. Resolution can be up to 1080p and bitrate can be up to 8.5 Mbps. Audio is transcoded only for renditions 360p and below; above that, audio is passed through.

BASIC: Amazon IVS delivers the original input to viewers. The viewer’s video-quality choice is limited to the original input. Resolution can be up to 480p and bitrate can be up to 1.5 Mbps.

channel_tags :: Lens' Channel (Maybe (HashMap Text Text))

Array of 1-50 maps, each of the form string:string (key:value).

channel_ingestEndpoint :: Lens' Channel (Maybe Text)

Channel ingest endpoint, part of the definition of an ingest server, used when you set up streaming software.

ChannelSummary

data ChannelSummary

Summary information about a channel.

See: newChannelSummary smart constructor.

Constructors

ChannelSummary'

Fields

authorized :: Maybe Bool

Whether the channel is private (enabled for playback authorization). Default: false.

arn :: Maybe Text

Channel ARN.

latencyMode :: Maybe ChannelLatencyMode

Channel latency mode. Use NORMAL to broadcast and deliver live video up to Full HD. Use LOW for near-real-time interaction with viewers. Default: LOW. (Note: In the Amazon IVS console, LOW and NORMAL correspond to Ultra-low and Standard, respectively.)

name :: Maybe Text

Channel name.

recordingConfigurationArn :: Maybe Text

Recording-configuration ARN. A value other than an empty string indicates that recording is enabled. Default: "" (empty string, recording is disabled).

tags :: Maybe (HashMap Text Text)

Array of 1-50 maps, each of the form string:string (key:value).

newChannelSummary :: ChannelSummary

Create a value of ChannelSummary with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:authorized:ChannelSummary', channelSummary_authorized - Whether the channel is private (enabled for playback authorization). Default: false.

$sel:arn:ChannelSummary', channelSummary_arn - Channel ARN.

$sel:latencyMode:ChannelSummary', channelSummary_latencyMode - Channel latency mode. Use NORMAL to broadcast and deliver live video up to Full HD. Use LOW for near-real-time interaction with viewers. Default: LOW. (Note: In the Amazon IVS console, LOW and NORMAL correspond to Ultra-low and Standard, respectively.)

$sel:name:ChannelSummary', channelSummary_name - Channel name.

$sel:recordingConfigurationArn:ChannelSummary', channelSummary_recordingConfigurationArn - Recording-configuration ARN. A value other than an empty string indicates that recording is enabled. Default: "" (empty string, recording is disabled).

$sel:tags:ChannelSummary', channelSummary_tags - Array of 1-50 maps, each of the form string:string (key:value).

channelSummary_authorized :: Lens' ChannelSummary (Maybe Bool)

Whether the channel is private (enabled for playback authorization). Default: false.

channelSummary_arn :: Lens' ChannelSummary (Maybe Text)

Channel ARN.

channelSummary_latencyMode :: Lens' ChannelSummary (Maybe ChannelLatencyMode)

Channel latency mode. Use NORMAL to broadcast and deliver live video up to Full HD. Use LOW for near-real-time interaction with viewers. Default: LOW. (Note: In the Amazon IVS console, LOW and NORMAL correspond to Ultra-low and Standard, respectively.)

channelSummary_name :: Lens' ChannelSummary (Maybe Text)

Channel name.

channelSummary_recordingConfigurationArn :: Lens' ChannelSummary (Maybe Text)

Recording-configuration ARN. A value other than an empty string indicates that recording is enabled. Default: "" (empty string, recording is disabled).

channelSummary_tags :: Lens' ChannelSummary (Maybe (HashMap Text Text))

Array of 1-50 maps, each of the form string:string (key:value).

DestinationConfiguration

data DestinationConfiguration

A complex type that describes a location where recorded videos will be stored. Each member represents a type of destination configuration. For recording, you define one and only one type of destination configuration.

See: newDestinationConfiguration smart constructor.

Constructors

DestinationConfiguration'

Fields

s3 :: Maybe S3DestinationConfiguration

An S3 destination configuration where recorded videos will be stored.

newDestinationConfiguration :: DestinationConfiguration

Create a value of DestinationConfiguration with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:s3:DestinationConfiguration', destinationConfiguration_s3 - An S3 destination configuration where recorded videos will be stored.

destinationConfiguration_s3 :: Lens' DestinationConfiguration (Maybe S3DestinationConfiguration)

An S3 destination configuration where recorded videos will be stored.

PlaybackKeyPair

data PlaybackKeyPair

A key pair used to sign and validate a playback authorization token.

See: newPlaybackKeyPair smart constructor.

Constructors

PlaybackKeyPair'

Fields

arn :: Maybe Text

Key-pair ARN.

fingerprint :: Maybe Text

Key-pair identifier.

name :: Maybe Text

Playback-key-pair name. The value does not need to be unique.

tags :: Maybe (HashMap Text Text)

Array of 1-50 maps, each of the form string:string (key:value).

newPlaybackKeyPair :: PlaybackKeyPair

Create a value of PlaybackKeyPair with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:arn:PlaybackKeyPair', playbackKeyPair_arn - Key-pair ARN.

$sel:fingerprint:PlaybackKeyPair', playbackKeyPair_fingerprint - Key-pair identifier.

$sel:name:PlaybackKeyPair', playbackKeyPair_name - Playback-key-pair name. The value does not need to be unique.

$sel:tags:PlaybackKeyPair', playbackKeyPair_tags - Array of 1-50 maps, each of the form string:string (key:value).

playbackKeyPair_arn :: Lens' PlaybackKeyPair (Maybe Text)

Key-pair ARN.

playbackKeyPair_fingerprint :: Lens' PlaybackKeyPair (Maybe Text)

Key-pair identifier.

playbackKeyPair_name :: Lens' PlaybackKeyPair (Maybe Text)

Playback-key-pair name. The value does not need to be unique.

playbackKeyPair_tags :: Lens' PlaybackKeyPair (Maybe (HashMap Text Text))

Array of 1-50 maps, each of the form string:string (key:value).

PlaybackKeyPairSummary

data PlaybackKeyPairSummary

Summary information about a playback key pair.

See: newPlaybackKeyPairSummary smart constructor.

Constructors

PlaybackKeyPairSummary'

Fields

arn :: Maybe Text

Key-pair ARN.

name :: Maybe Text

Playback-key-pair name. The value does not need to be unique.

tags :: Maybe (HashMap Text Text)

Array of 1-50 maps, each of the form string:string (key:value).

newPlaybackKeyPairSummary :: PlaybackKeyPairSummary

Create a value of PlaybackKeyPairSummary with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:arn:PlaybackKeyPairSummary', playbackKeyPairSummary_arn - Key-pair ARN.

$sel:name:PlaybackKeyPairSummary', playbackKeyPairSummary_name - Playback-key-pair name. The value does not need to be unique.

$sel:tags:PlaybackKeyPairSummary', playbackKeyPairSummary_tags - Array of 1-50 maps, each of the form string:string (key:value).

playbackKeyPairSummary_arn :: Lens' PlaybackKeyPairSummary (Maybe Text)

Key-pair ARN.

playbackKeyPairSummary_name :: Lens' PlaybackKeyPairSummary (Maybe Text)

Playback-key-pair name. The value does not need to be unique.

playbackKeyPairSummary_tags :: Lens' PlaybackKeyPairSummary (Maybe (HashMap Text Text))

Array of 1-50 maps, each of the form string:string (key:value).

RecordingConfiguration

data RecordingConfiguration

An object representing a configuration to record a channel stream.

See: newRecordingConfiguration smart constructor.

Constructors

RecordingConfiguration'

Fields

name :: Maybe Text

Recording-configuration name. The value does not need to be unique.

tags :: Maybe (HashMap Text Text)

Array of 1-50 maps, each of the form string:string (key:value).

arn :: Text

Recording-configuration ARN.

destinationConfiguration :: DestinationConfiguration

A complex type that contains information about where recorded video will be stored.

state :: RecordingConfigurationState

Indicates the current state of the recording configuration. When the state is ACTIVE, the configuration is ready for recording a channel stream.

newRecordingConfiguration

Arguments

:: Text
$sel:arn:RecordingConfiguration'

-> DestinationConfiguration
$sel:destinationConfiguration:RecordingConfiguration'

-> RecordingConfigurationState
$sel:state:RecordingConfiguration'

-> RecordingConfiguration

Create a value of RecordingConfiguration with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:name:RecordingConfiguration', recordingConfiguration_name - Recording-configuration name. The value does not need to be unique.

$sel:tags:RecordingConfiguration', recordingConfiguration_tags - Array of 1-50 maps, each of the form string:string (key:value).

$sel:arn:RecordingConfiguration', recordingConfiguration_arn - Recording-configuration ARN.

$sel:destinationConfiguration:RecordingConfiguration', recordingConfiguration_destinationConfiguration - A complex type that contains information about where recorded video will be stored.

$sel:state:RecordingConfiguration', recordingConfiguration_state - Indicates the current state of the recording configuration. When the state is ACTIVE, the configuration is ready for recording a channel stream.

recordingConfiguration_name :: Lens' RecordingConfiguration (Maybe Text)

Recording-configuration name. The value does not need to be unique.

recordingConfiguration_tags :: Lens' RecordingConfiguration (Maybe (HashMap Text Text))

Array of 1-50 maps, each of the form string:string (key:value).

recordingConfiguration_arn :: Lens' RecordingConfiguration Text

Recording-configuration ARN.

recordingConfiguration_destinationConfiguration :: Lens' RecordingConfiguration DestinationConfiguration

A complex type that contains information about where recorded video will be stored.

recordingConfiguration_state :: Lens' RecordingConfiguration RecordingConfigurationState

Indicates the current state of the recording configuration. When the state is ACTIVE, the configuration is ready for recording a channel stream.

RecordingConfigurationSummary

data RecordingConfigurationSummary

Summary information about a RecordingConfiguration.

See: newRecordingConfigurationSummary smart constructor.

Constructors

RecordingConfigurationSummary'

Fields

name :: Maybe Text

Recording-configuration name. The value does not need to be unique.

tags :: Maybe (HashMap Text Text)

Array of 1-50 maps, each of the form string:string (key:value).

arn :: Text

Recording-configuration ARN.

destinationConfiguration :: DestinationConfiguration

A complex type that contains information about where recorded video will be stored.

state :: RecordingConfigurationState

Indicates the current state of the recording configuration. When the state is ACTIVE, the configuration is ready for recording a channel stream.

newRecordingConfigurationSummary

Arguments

:: Text
$sel:arn:RecordingConfigurationSummary'

-> DestinationConfiguration
$sel:destinationConfiguration:RecordingConfigurationSummary'

-> RecordingConfigurationState
$sel:state:RecordingConfigurationSummary'

-> RecordingConfigurationSummary

Create a value of RecordingConfigurationSummary with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:name:RecordingConfigurationSummary', recordingConfigurationSummary_name - Recording-configuration name. The value does not need to be unique.

$sel:tags:RecordingConfigurationSummary', recordingConfigurationSummary_tags - Array of 1-50 maps, each of the form string:string (key:value).

$sel:arn:RecordingConfigurationSummary', recordingConfigurationSummary_arn - Recording-configuration ARN.

$sel:destinationConfiguration:RecordingConfigurationSummary', recordingConfigurationSummary_destinationConfiguration - A complex type that contains information about where recorded video will be stored.

$sel:state:RecordingConfigurationSummary', recordingConfigurationSummary_state - Indicates the current state of the recording configuration. When the state is ACTIVE, the configuration is ready for recording a channel stream.

recordingConfigurationSummary_name :: Lens' RecordingConfigurationSummary (Maybe Text)

Recording-configuration name. The value does not need to be unique.

recordingConfigurationSummary_tags :: Lens' RecordingConfigurationSummary (Maybe (HashMap Text Text))

Array of 1-50 maps, each of the form string:string (key:value).

recordingConfigurationSummary_arn :: Lens' RecordingConfigurationSummary Text

Recording-configuration ARN.

recordingConfigurationSummary_destinationConfiguration :: Lens' RecordingConfigurationSummary DestinationConfiguration

A complex type that contains information about where recorded video will be stored.

recordingConfigurationSummary_state :: Lens' RecordingConfigurationSummary RecordingConfigurationState

Indicates the current state of the recording configuration. When the state is ACTIVE, the configuration is ready for recording a channel stream.

S3DestinationConfiguration

data S3DestinationConfiguration

A complex type that describes an S3 location where recorded videos will be stored.

See: newS3DestinationConfiguration smart constructor.

Constructors

S3DestinationConfiguration'

Fields

bucketName :: Text

Location (S3 bucket name) where recorded videos will be stored.

newS3DestinationConfiguration

Arguments

:: Text
$sel:bucketName:S3DestinationConfiguration'

-> S3DestinationConfiguration

Create a value of S3DestinationConfiguration with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:bucketName:S3DestinationConfiguration', s3DestinationConfiguration_bucketName - Location (S3 bucket name) where recorded videos will be stored.

s3DestinationConfiguration_bucketName :: Lens' S3DestinationConfiguration Text

Location (S3 bucket name) where recorded videos will be stored.

Stream

data Stream

Specifies a live video stream that has been ingested and distributed.

See: newStream smart constructor.

Constructors

Stream'

Fields

playbackUrl :: Maybe Text

URL of the master playlist, required by the video player to play the HLS stream.

state :: Maybe StreamState

The stream’s state.

startTime :: Maybe POSIX

Time of the stream’s start. This is an ISO 8601 timestamp returned as a string.

channelArn :: Maybe Text

Channel ARN for the stream.

viewerCount :: Maybe Integer

A count of concurrent views of the stream. Typically, a new view appears in viewerCount within 15 seconds of when video playback starts and a view is removed from viewerCount within 1 minute of when video playback ends. A value of -1 indicates that the request timed out; in this case, retry.

health :: Maybe StreamHealth

The stream’s health.

newStream :: Stream

Create a value of Stream with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:playbackUrl:Stream', stream_playbackUrl - URL of the master playlist, required by the video player to play the HLS stream.

$sel:state:Stream', stream_state - The stream’s state.

$sel:startTime:Stream', stream_startTime - Time of the stream’s start. This is an ISO 8601 timestamp returned as a string.

$sel:channelArn:Stream', stream_channelArn - Channel ARN for the stream.

$sel:viewerCount:Stream', stream_viewerCount - A count of concurrent views of the stream. Typically, a new view appears in viewerCount within 15 seconds of when video playback starts and a view is removed from viewerCount within 1 minute of when video playback ends. A value of -1 indicates that the request timed out; in this case, retry.

$sel:health:Stream', stream_health - The stream’s health.

stream_playbackUrl :: Lens' Stream (Maybe Text)

URL of the master playlist, required by the video player to play the HLS stream.

stream_state :: Lens' Stream (Maybe StreamState)

The stream’s state.

stream_startTime :: Lens' Stream (Maybe UTCTime)

Time of the stream’s start. This is an ISO 8601 timestamp returned as a string.

stream_channelArn :: Lens' Stream (Maybe Text)

Channel ARN for the stream.

stream_viewerCount :: Lens' Stream (Maybe Integer)

A count of concurrent views of the stream. Typically, a new view appears in viewerCount within 15 seconds of when video playback starts and a view is removed from viewerCount within 1 minute of when video playback ends. A value of -1 indicates that the request timed out; in this case, retry.

stream_health :: Lens' Stream (Maybe StreamHealth)

The stream’s health.

StreamKey

data StreamKey

Object specifying a stream key.

See: newStreamKey smart constructor.

Constructors

StreamKey'

Fields

arn :: Maybe Text

Stream-key ARN.

value :: Maybe (Sensitive Text)

Stream-key value.

channelArn :: Maybe Text

Channel ARN for the stream.

tags :: Maybe (HashMap Text Text)

Array of 1-50 maps, each of the form string:string (key:value).

newStreamKey :: StreamKey

Create a value of StreamKey with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:arn:StreamKey', streamKey_arn - Stream-key ARN.

$sel:value:StreamKey', streamKey_value - Stream-key value.

$sel:channelArn:StreamKey', streamKey_channelArn - Channel ARN for the stream.

$sel:tags:StreamKey', streamKey_tags - Array of 1-50 maps, each of the form string:string (key:value).

streamKey_arn :: Lens' StreamKey (Maybe Text)

Stream-key ARN.

streamKey_value :: Lens' StreamKey (Maybe Text)

Stream-key value.

streamKey_channelArn :: Lens' StreamKey (Maybe Text)

Channel ARN for the stream.

streamKey_tags :: Lens' StreamKey (Maybe (HashMap Text Text))

Array of 1-50 maps, each of the form string:string (key:value).

StreamKeySummary

data StreamKeySummary

Summary information about a stream key.

See: newStreamKeySummary smart constructor.

Constructors

StreamKeySummary'

Fields

arn :: Maybe Text

Stream-key ARN.

channelArn :: Maybe Text

Channel ARN for the stream.

tags :: Maybe (HashMap Text Text)

Array of 1-50 maps, each of the form string:string (key:value).

newStreamKeySummary :: StreamKeySummary

Create a value of StreamKeySummary with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:arn:StreamKeySummary', streamKeySummary_arn - Stream-key ARN.

$sel:channelArn:StreamKeySummary', streamKeySummary_channelArn - Channel ARN for the stream.

$sel:tags:StreamKeySummary', streamKeySummary_tags - Array of 1-50 maps, each of the form string:string (key:value).

streamKeySummary_arn :: Lens' StreamKeySummary (Maybe Text)

Stream-key ARN.

streamKeySummary_channelArn :: Lens' StreamKeySummary (Maybe Text)

Channel ARN for the stream.

streamKeySummary_tags :: Lens' StreamKeySummary (Maybe (HashMap Text Text))

Array of 1-50 maps, each of the form string:string (key:value).

StreamSummary

data StreamSummary

Summary information about a stream.

See: newStreamSummary smart constructor.

Constructors

StreamSummary'

Fields

state :: Maybe StreamState

The stream’s state.

startTime :: Maybe POSIX

Time of the stream’s start. This is an ISO 8601 timestamp returned as a string.

channelArn :: Maybe Text

Channel ARN for the stream.

viewerCount :: Maybe Integer

A count of concurrent views of the stream. Typically, a new view appears in viewerCount within 15 seconds of when video playback starts and a view is removed from viewerCount within 1 minute of when video playback ends. A value of -1 indicates that the request timed out; in this case, retry.

health :: Maybe StreamHealth

The stream’s health.

newStreamSummary :: StreamSummary

Create a value of StreamSummary with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:state:StreamSummary', streamSummary_state - The stream’s state.

$sel:startTime:StreamSummary', streamSummary_startTime - Time of the stream’s start. This is an ISO 8601 timestamp returned as a string.

$sel:channelArn:StreamSummary', streamSummary_channelArn - Channel ARN for the stream.

$sel:viewerCount:StreamSummary', streamSummary_viewerCount - A count of concurrent views of the stream. Typically, a new view appears in viewerCount within 15 seconds of when video playback starts and a view is removed from viewerCount within 1 minute of when video playback ends. A value of -1 indicates that the request timed out; in this case, retry.

$sel:health:StreamSummary', streamSummary_health - The stream’s health.

streamSummary_state :: Lens' StreamSummary (Maybe StreamState)

The stream’s state.

streamSummary_startTime :: Lens' StreamSummary (Maybe UTCTime)

Time of the stream’s start. This is an ISO 8601 timestamp returned as a string.

streamSummary_channelArn :: Lens' StreamSummary (Maybe Text)

Channel ARN for the stream.

streamSummary_viewerCount :: Lens' StreamSummary (Maybe Integer)

A count of concurrent views of the stream. Typically, a new view appears in viewerCount within 15 seconds of when video playback starts and a view is removed from viewerCount within 1 minute of when video playback ends. A value of -1 indicates that the request timed out; in this case, retry.

streamSummary_health :: Lens' StreamSummary (Maybe StreamHealth)

The stream’s health.