Amazonka.IVS.Types.ChannelSummary

Description

Documentation

data ChannelSummary

Summary information about a channel.

See: newChannelSummary smart constructor.

Constructors

ChannelSummary'

Fields

authorized :: Maybe Bool

Whether the channel is private (enabled for playback authorization). Default: false.

arn :: Maybe Text

Channel ARN.

latencyMode :: Maybe ChannelLatencyMode

Channel latency mode. Use NORMAL to broadcast and deliver live video up to Full HD. Use LOW for near-real-time interaction with viewers. Default: LOW. (Note: In the Amazon IVS console, LOW and NORMAL correspond to Ultra-low and Standard, respectively.)

name :: Maybe Text

Channel name.

recordingConfigurationArn :: Maybe Text

Recording-configuration ARN. A value other than an empty string indicates that recording is enabled. Default: "" (empty string, recording is disabled).

tags :: Maybe (HashMap Text Text)

Array of 1-50 maps, each of the form string:string (key:value).

newChannelSummary :: ChannelSummary

Create a value of ChannelSummary with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:authorized:ChannelSummary', channelSummary_authorized - Whether the channel is private (enabled for playback authorization). Default: false.

$sel:arn:ChannelSummary', channelSummary_arn - Channel ARN.

$sel:latencyMode:ChannelSummary', channelSummary_latencyMode - Channel latency mode. Use NORMAL to broadcast and deliver live video up to Full HD. Use LOW for near-real-time interaction with viewers. Default: LOW. (Note: In the Amazon IVS console, LOW and NORMAL correspond to Ultra-low and Standard, respectively.)

$sel:name:ChannelSummary', channelSummary_name - Channel name.

$sel:recordingConfigurationArn:ChannelSummary', channelSummary_recordingConfigurationArn - Recording-configuration ARN. A value other than an empty string indicates that recording is enabled. Default: "" (empty string, recording is disabled).

$sel:tags:ChannelSummary', channelSummary_tags - Array of 1-50 maps, each of the form string:string (key:value).

channelSummary_authorized :: Lens' ChannelSummary (Maybe Bool)

Whether the channel is private (enabled for playback authorization). Default: false.

channelSummary_arn :: Lens' ChannelSummary (Maybe Text)

Channel ARN.

channelSummary_latencyMode :: Lens' ChannelSummary (Maybe ChannelLatencyMode)

Channel latency mode. Use NORMAL to broadcast and deliver live video up to Full HD. Use LOW for near-real-time interaction with viewers. Default: LOW. (Note: In the Amazon IVS console, LOW and NORMAL correspond to Ultra-low and Standard, respectively.)

channelSummary_name :: Lens' ChannelSummary (Maybe Text)

Channel name.

channelSummary_recordingConfigurationArn :: Lens' ChannelSummary (Maybe Text)

Recording-configuration ARN. A value other than an empty string indicates that recording is enabled. Default: "" (empty string, recording is disabled).

channelSummary_tags :: Lens' ChannelSummary (Maybe (HashMap Text Text))

Array of 1-50 maps, each of the form string:string (key:value).