Amazonka.IdentityStore

Description

Derived from API version 2020-06-15 of the AWS service descriptions, licensed under Apache 2.0.

The AWS Single Sign-On (SSO) Identity Store service provides a single place to retrieve all of your identities (users and groups). For more information about AWS, see the AWS Single Sign-On User Guide.

Service Configuration

defaultService :: Service

API version 2020-06-15 of the Amazon SSO Identity Store SDK configuration.

Errors

Error matchers are designed for use with the functions provided by Control.Exception.Lens. This allows catching (and rethrowing) service specific errors returned by IdentityStore.

ValidationException

_ValidationException :: AsError a => Getting (First ServiceError) a ServiceError

The request failed because it contains a syntax error.

AccessDeniedException

_AccessDeniedException :: AsError a => Getting (First ServiceError) a ServiceError

You do not have sufficient access to perform this action.

ThrottlingException

_ThrottlingException :: AsError a => Getting (First ServiceError) a ServiceError

Indicates that the principal has crossed the throttling limits of the API operations.

InternalServerException

_InternalServerException :: AsError a => Getting (First ServiceError) a ServiceError

The request processing has failed because of an unknown error, exception or failure with an internal server.

ResourceNotFoundException

_ResourceNotFoundException :: AsError a => Getting (First ServiceError) a ServiceError

Indicates that a requested resource is not found.

Waiters

Waiters poll by repeatedly sending a request until some remote success condition configured by the Wait specification is fulfilled. The Wait specification determines how many attempts should be made, in addition to delay and retry strategies.

Operations

Some AWS operations return results that are incomplete and require subsequent requests in order to obtain the entire result set. The process of sending subsequent requests to continue where a previous request left off is called pagination. For example, the ListObjects operation of Amazon S3 returns up to 1000 objects at a time, and you must send subsequent requests with the appropriate Marker in order to retrieve the next page of results.

Operations that have an AWSPager instance can transparently perform subsequent requests, correctly setting Markers and other request facets to iterate through the entire result set of a truncated API operation. Operations which support this have an additional note in the documentation.

Many operations have the ability to filter results on the server side. See the individual operation parameters for details.

DescribeGroup

data DescribeGroup

See: newDescribeGroup smart constructor.

Constructors

DescribeGroup' Text Text

newDescribeGroup

Arguments

:: Text
$sel:identityStoreId:DescribeGroup'

-> Text
$sel:groupId:DescribeGroup'

-> DescribeGroup

Create a value of DescribeGroup with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:identityStoreId:DescribeGroup', describeGroup_identityStoreId - The globally unique identifier for the identity store, such as d-1234567890. In this example, d- is a fixed prefix, and 1234567890 is a randomly generated string that contains number and lower case letters. This value is generated at the time that a new identity store is created.

$sel:groupId:DescribeGroup', describeGroup_groupId - The identifier for a group in the identity store.

data DescribeGroupResponse

See: newDescribeGroupResponse smart constructor.

Constructors

DescribeGroupResponse' Int Text Text

newDescribeGroupResponse

Arguments

:: Int
$sel:httpStatus:DescribeGroupResponse'

-> Text
$sel:groupId:DescribeGroup'

-> Text
$sel:displayName:DescribeGroupResponse'

-> DescribeGroupResponse

Create a value of DescribeGroupResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:httpStatus:DescribeGroupResponse', describeGroupResponse_httpStatus - The response's http status code.

$sel:groupId:DescribeGroup', describeGroupResponse_groupId - The identifier for a group in the identity store.

$sel:displayName:DescribeGroupResponse', describeGroupResponse_displayName - Contains the group’s display name value. The length limit is 1,024 characters. This value can consist of letters, accented characters, symbols, numbers, punctuation, tab, new line, carriage return, space, and nonbreaking space in this attribute. The characters <>;:% are excluded. This value is specified at the time that the group is created and stored as an attribute of the group object in the identity store.

ListUsers

data ListUsers

See: newListUsers smart constructor.

Constructors

ListUsers' (Maybe [Filter]) (Maybe Text) (Maybe Natural) Text

newListUsers

Arguments

:: Text
$sel:identityStoreId:ListUsers'

-> ListUsers

Create a value of ListUsers with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:filters:ListUsers', listUsers_filters - A list of Filter objects, which is used in the ListUsers and ListGroups request.

$sel:nextToken:ListUsers', listUsers_nextToken - The pagination token used for the ListUsers and ListGroups API operations. This value is generated by the identity store service. It is returned in the API response if the total results are more than the size of one page. This token is also returned when it is used in the API request to search for the next page.

$sel:maxResults:ListUsers', listUsers_maxResults - The maximum number of results to be returned per request. This parameter is used in the ListUsers and ListGroups request to specify how many results to return in one page. The length limit is 50 characters.

$sel:identityStoreId:ListUsers', listUsers_identityStoreId - The globally unique identifier for the identity store, such as d-1234567890. In this example, d- is a fixed prefix, and 1234567890 is a randomly generated string that contains number and lower case letters. This value is generated at the time that a new identity store is created.

data ListUsersResponse

See: newListUsersResponse smart constructor.

Constructors

ListUsersResponse' (Maybe Text) Int [User]

newListUsersResponse

Arguments

:: Int
$sel:httpStatus:ListUsersResponse'

-> ListUsersResponse

Create a value of ListUsersResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:nextToken:ListUsers', listUsersResponse_nextToken - The pagination token used for the ListUsers and ListGroups API operations. This value is generated by the identity store service. It is returned in the API response if the total results are more than the size of one page. This token is also returned when it is used in the API request to search for the next page.

$sel:httpStatus:ListUsersResponse', listUsersResponse_httpStatus - The response's http status code.

$sel:users:ListUsersResponse', listUsersResponse_users - A list of User objects in the identity store.

DescribeUser

data DescribeUser

See: newDescribeUser smart constructor.

Constructors

DescribeUser' Text Text

newDescribeUser

Arguments

:: Text
$sel:identityStoreId:DescribeUser'

-> Text
$sel:userId:DescribeUser'

-> DescribeUser

Create a value of DescribeUser with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:identityStoreId:DescribeUser', describeUser_identityStoreId - The globally unique identifier for the identity store, such as d-1234567890. In this example, d- is a fixed prefix, and 1234567890 is a randomly generated string that contains number and lower case letters. This value is generated at the time that a new identity store is created.

$sel:userId:DescribeUser', describeUser_userId - The identifier for a user in the identity store.

data DescribeUserResponse

See: newDescribeUserResponse smart constructor.

Constructors

DescribeUserResponse' Int (Sensitive Text) Text

newDescribeUserResponse

Arguments

:: Int
$sel:httpStatus:DescribeUserResponse'

-> Text
$sel:userName:DescribeUserResponse'

-> Text
$sel:userId:DescribeUser'

-> DescribeUserResponse

Create a value of DescribeUserResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:httpStatus:DescribeUserResponse', describeUserResponse_httpStatus - The response's http status code.

$sel:userName:DescribeUserResponse', describeUserResponse_userName - Contains the user’s user name value. The length limit is 128 characters. This value can consist of letters, accented characters, symbols, numbers, and punctuation. The characters <>;:% are excluded. This value is specified at the time the user is created and stored as an attribute of the user object in the identity store.

$sel:userId:DescribeUser', describeUserResponse_userId - The identifier for a user in the identity store.

ListGroups

data ListGroups

See: newListGroups smart constructor.

Constructors

ListGroups' (Maybe [Filter]) (Maybe Text) (Maybe Natural) Text

newListGroups

Arguments

:: Text
$sel:identityStoreId:ListGroups'

-> ListGroups

Create a value of ListGroups with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:filters:ListGroups', listGroups_filters - A list of Filter objects, which is used in the ListUsers and ListGroups request.

$sel:nextToken:ListGroups', listGroups_nextToken - The pagination token used for the ListUsers and ListGroups API operations. This value is generated by the identity store service. It is returned in the API response if the total results are more than the size of one page. This token is also returned when it is used in the API request to search for the next page.

$sel:maxResults:ListGroups', listGroups_maxResults - The maximum number of results to be returned per request. This parameter is used in the ListUsers and ListGroups request to specify how many results to return in one page. The length limit is 50 characters.

$sel:identityStoreId:ListGroups', listGroups_identityStoreId - The globally unique identifier for the identity store, such as d-1234567890. In this example, d- is a fixed prefix, and 1234567890 is a randomly generated string that contains number and lower case letters. This value is generated at the time that a new identity store is created.

data ListGroupsResponse

See: newListGroupsResponse smart constructor.

Constructors

ListGroupsResponse' (Maybe Text) Int [Group]

newListGroupsResponse

Arguments

:: Int
$sel:httpStatus:ListGroupsResponse'

-> ListGroupsResponse

Create a value of ListGroupsResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:nextToken:ListGroups', listGroupsResponse_nextToken - The pagination token used for the ListUsers and ListGroups API operations. This value is generated by the identity store service. It is returned in the API response if the total results are more than the size of one page. This token is also returned when it1 is used in the API request to search for the next page.

$sel:httpStatus:ListGroupsResponse', listGroupsResponse_httpStatus - The response's http status code.

$sel:groups:ListGroupsResponse', listGroupsResponse_groups - A list of Group objects in the identity store.

Types

Filter

data Filter

A query filter used by ListUsers and ListGroup. This filter object provides the attribute name and attribute value to search users or groups.

See: newFilter smart constructor.

Constructors

Filter' Text (Sensitive Text)

newFilter

Arguments

:: Text
$sel:attributePath:Filter'

-> Text
$sel:attributeValue:Filter'

-> Filter

Create a value of Filter with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:attributePath:Filter', filter_attributePath - The attribute path that is used to specify which attribute name to search. Length limit is 255 characters. For example, UserName is a valid attribute path for the ListUsers API, and DisplayName is a valid attribute path for the ListGroups API.

$sel:attributeValue:Filter', filter_attributeValue - Represents the data for an attribute. Each attribute value is described as a name-value pair.

Group

data Group

A group object, which contains a specified group’s metadata and attributes.

See: newGroup smart constructor.

Constructors

Group' Text Text

newGroup

Arguments

:: Text
$sel:groupId:Group'

-> Text
$sel:displayName:Group'

-> Group

Create a value of Group with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:groupId:Group', group_groupId - The identifier for a group in the identity store.

$sel:displayName:Group', group_displayName - Contains the group’s display name value. The length limit is 1,024 characters. This value can consist of letters, accented characters, symbols, numbers, punctuation, tab, new line, carriage return, space, and nonbreaking space in this attribute. The characters <>;:% are excluded. This value is specified at the time the group is created and stored as an attribute of the group object in the identity store.

User

data User

A user object, which contains a specified user’s metadata and attributes.

See: newUser smart constructor.

Constructors

User' (Sensitive Text) Text

newUser

Arguments

:: Text
$sel:userName:User'

-> Text
$sel:userId:User'

-> User

Create a value of User with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:userName:User', user_userName - Contains the user’s user name value. The length limit is 128 characters. This value can consist of letters, accented characters, symbols, numbers, and punctuation. The characters <>;:% are excluded. This value is specified at the time the user is created and stored as an attribute of the user object in the identity store.

$sel:userId:User', user_userId - The identifier for a user in the identity store.