Amazonka.ImageBuilder.CreateImagePipeline

Description

Creates a new image pipeline. Image pipelines enable you to automate the creation and distribution of images.

Creating a Request

data CreateImagePipeline

See: newCreateImagePipeline smart constructor.

Constructors

CreateImagePipeline'

Fields

status :: Maybe PipelineStatus

The status of the image pipeline.

containerRecipeArn :: Maybe Text

The Amazon Resource Name (ARN) of the container recipe that is used to configure images created by this container pipeline.

imageTestsConfiguration :: Maybe ImageTestsConfiguration

The image test configuration of the image pipeline.

schedule :: Maybe Schedule

The schedule of the image pipeline.

enhancedImageMetadataEnabled :: Maybe Bool

Collects additional information about the image being created, including the operating system (OS) version and package list. This information is used to enhance the overall experience of using EC2 Image Builder. Enabled by default.

distributionConfigurationArn :: Maybe Text

The Amazon Resource Name (ARN) of the distribution configuration that will be used to configure and distribute images created by this image pipeline.

imageRecipeArn :: Maybe Text

The Amazon Resource Name (ARN) of the image recipe that will be used to configure images created by this image pipeline.

description :: Maybe Text

The description of the image pipeline.

tags :: Maybe (HashMap Text Text)

The tags of the image pipeline.

name :: Text

The name of the image pipeline.

infrastructureConfigurationArn :: Text

The Amazon Resource Name (ARN) of the infrastructure configuration that will be used to build images created by this image pipeline.

clientToken :: Text

The idempotency token used to make this request idempotent.

newCreateImagePipeline

Arguments

:: Text
$sel:name:CreateImagePipeline'

-> Text
$sel:infrastructureConfigurationArn:CreateImagePipeline'

-> Text
$sel:clientToken:CreateImagePipeline'

-> CreateImagePipeline

Create a value of CreateImagePipeline with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:status:CreateImagePipeline', createImagePipeline_status - The status of the image pipeline.

$sel:containerRecipeArn:CreateImagePipeline', createImagePipeline_containerRecipeArn - The Amazon Resource Name (ARN) of the container recipe that is used to configure images created by this container pipeline.

$sel:imageTestsConfiguration:CreateImagePipeline', createImagePipeline_imageTestsConfiguration - The image test configuration of the image pipeline.

$sel:schedule:CreateImagePipeline', createImagePipeline_schedule - The schedule of the image pipeline.

$sel:enhancedImageMetadataEnabled:CreateImagePipeline', createImagePipeline_enhancedImageMetadataEnabled - Collects additional information about the image being created, including the operating system (OS) version and package list. This information is used to enhance the overall experience of using EC2 Image Builder. Enabled by default.

$sel:distributionConfigurationArn:CreateImagePipeline', createImagePipeline_distributionConfigurationArn - The Amazon Resource Name (ARN) of the distribution configuration that will be used to configure and distribute images created by this image pipeline.

$sel:imageRecipeArn:CreateImagePipeline', createImagePipeline_imageRecipeArn - The Amazon Resource Name (ARN) of the image recipe that will be used to configure images created by this image pipeline.

$sel:description:CreateImagePipeline', createImagePipeline_description - The description of the image pipeline.

$sel:tags:CreateImagePipeline', createImagePipeline_tags - The tags of the image pipeline.

$sel:name:CreateImagePipeline', createImagePipeline_name - The name of the image pipeline.

$sel:infrastructureConfigurationArn:CreateImagePipeline', createImagePipeline_infrastructureConfigurationArn - The Amazon Resource Name (ARN) of the infrastructure configuration that will be used to build images created by this image pipeline.

$sel:clientToken:CreateImagePipeline', createImagePipeline_clientToken - The idempotency token used to make this request idempotent.

Request Lenses

createImagePipeline_status :: Lens' CreateImagePipeline (Maybe PipelineStatus)

The status of the image pipeline.

createImagePipeline_containerRecipeArn :: Lens' CreateImagePipeline (Maybe Text)

The Amazon Resource Name (ARN) of the container recipe that is used to configure images created by this container pipeline.

createImagePipeline_imageTestsConfiguration :: Lens' CreateImagePipeline (Maybe ImageTestsConfiguration)

The image test configuration of the image pipeline.

createImagePipeline_schedule :: Lens' CreateImagePipeline (Maybe Schedule)

The schedule of the image pipeline.

createImagePipeline_enhancedImageMetadataEnabled :: Lens' CreateImagePipeline (Maybe Bool)

Collects additional information about the image being created, including the operating system (OS) version and package list. This information is used to enhance the overall experience of using EC2 Image Builder. Enabled by default.

createImagePipeline_distributionConfigurationArn :: Lens' CreateImagePipeline (Maybe Text)

The Amazon Resource Name (ARN) of the distribution configuration that will be used to configure and distribute images created by this image pipeline.

createImagePipeline_imageRecipeArn :: Lens' CreateImagePipeline (Maybe Text)

The Amazon Resource Name (ARN) of the image recipe that will be used to configure images created by this image pipeline.

createImagePipeline_description :: Lens' CreateImagePipeline (Maybe Text)

The description of the image pipeline.

createImagePipeline_tags :: Lens' CreateImagePipeline (Maybe (HashMap Text Text))

The tags of the image pipeline.

createImagePipeline_name :: Lens' CreateImagePipeline Text

The name of the image pipeline.

createImagePipeline_infrastructureConfigurationArn :: Lens' CreateImagePipeline Text

The Amazon Resource Name (ARN) of the infrastructure configuration that will be used to build images created by this image pipeline.

createImagePipeline_clientToken :: Lens' CreateImagePipeline Text

The idempotency token used to make this request idempotent.

Destructuring the Response

data CreateImagePipelineResponse

See: newCreateImagePipelineResponse smart constructor.

Constructors

CreateImagePipelineResponse'

Fields

requestId :: Maybe Text

The request ID that uniquely identifies this request.

clientToken :: Maybe Text

The idempotency token used to make this request idempotent.

imagePipelineArn :: Maybe Text

The Amazon Resource Name (ARN) of the image pipeline that was created by this request.

httpStatus :: Int

The response's http status code.

newCreateImagePipelineResponse

Arguments

:: Int
$sel:httpStatus:CreateImagePipelineResponse'

-> CreateImagePipelineResponse

Create a value of CreateImagePipelineResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:requestId:CreateImagePipelineResponse', createImagePipelineResponse_requestId - The request ID that uniquely identifies this request.

$sel:clientToken:CreateImagePipeline', createImagePipelineResponse_clientToken - The idempotency token used to make this request idempotent.

$sel:imagePipelineArn:CreateImagePipelineResponse', createImagePipelineResponse_imagePipelineArn - The Amazon Resource Name (ARN) of the image pipeline that was created by this request.

$sel:httpStatus:CreateImagePipelineResponse', createImagePipelineResponse_httpStatus - The response's http status code.

Response Lenses

createImagePipelineResponse_requestId :: Lens' CreateImagePipelineResponse (Maybe Text)

The request ID that uniquely identifies this request.

createImagePipelineResponse_clientToken :: Lens' CreateImagePipelineResponse (Maybe Text)

The idempotency token used to make this request idempotent.

createImagePipelineResponse_imagePipelineArn :: Lens' CreateImagePipelineResponse (Maybe Text)

The Amazon Resource Name (ARN) of the image pipeline that was created by this request.

createImagePipelineResponse_httpStatus :: Lens' CreateImagePipelineResponse Int

The response's http status code.