Amazonka.ImageBuilder.DeleteImage

Description

Deletes an Image Builder image resource. This does not delete any EC2 AMIs or ECR container images that are created during the image build process. You must clean those up separately, using the appropriate Amazon EC2 or Amazon ECR console actions, or API or CLI commands.

To deregister an EC2 Linux AMI, see Deregister your Linux AMI in the /Amazon EC2 User Guide/ .

To deregister an EC2 Windows AMI, see Deregister your Windows AMI in the /Amazon EC2 Windows Guide/ .

To delete a container image from Amazon ECR, see Deleting an image in the Amazon ECR User Guide.

Creating a Request

data DeleteImage

See: newDeleteImage smart constructor.

Constructors

DeleteImage'

Fields

imageBuildVersionArn :: Text

The Amazon Resource Name (ARN) of the Image Builder image resource to delete.

newDeleteImage

Arguments

:: Text
$sel:imageBuildVersionArn:DeleteImage'

-> DeleteImage

Create a value of DeleteImage with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:imageBuildVersionArn:DeleteImage', deleteImage_imageBuildVersionArn - The Amazon Resource Name (ARN) of the Image Builder image resource to delete.

Request Lenses

deleteImage_imageBuildVersionArn :: Lens' DeleteImage Text

The Amazon Resource Name (ARN) of the Image Builder image resource to delete.

Destructuring the Response

data DeleteImageResponse

See: newDeleteImageResponse smart constructor.

Constructors

DeleteImageResponse'

Fields

requestId :: Maybe Text

The request ID that uniquely identifies this request.

imageBuildVersionArn :: Maybe Text

The Amazon Resource Name (ARN) of the Image Builder image resource that was deleted.

httpStatus :: Int

The response's http status code.

newDeleteImageResponse

Arguments

:: Int
$sel:httpStatus:DeleteImageResponse'

-> DeleteImageResponse

Create a value of DeleteImageResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:requestId:DeleteImageResponse', deleteImageResponse_requestId - The request ID that uniquely identifies this request.

$sel:imageBuildVersionArn:DeleteImage', deleteImageResponse_imageBuildVersionArn - The Amazon Resource Name (ARN) of the Image Builder image resource that was deleted.

$sel:httpStatus:DeleteImageResponse', deleteImageResponse_httpStatus - The response's http status code.

Response Lenses

deleteImageResponse_requestId :: Lens' DeleteImageResponse (Maybe Text)

The request ID that uniquely identifies this request.

deleteImageResponse_imageBuildVersionArn :: Lens' DeleteImageResponse (Maybe Text)

The Amazon Resource Name (ARN) of the Image Builder image resource that was deleted.

deleteImageResponse_httpStatus :: Lens' DeleteImageResponse Int

The response's http status code.