Amazonka.ImageBuilder.GetContainerRecipePolicy

Description

Retrieves the policy for a container recipe.

Creating a Request

data GetContainerRecipePolicy

See: newGetContainerRecipePolicy smart constructor.

Constructors

GetContainerRecipePolicy'

Fields

containerRecipeArn :: Text

The Amazon Resource Name (ARN) of the container recipe for the policy being requested.

newGetContainerRecipePolicy

Arguments

:: Text
$sel:containerRecipeArn:GetContainerRecipePolicy'

-> GetContainerRecipePolicy

Create a value of GetContainerRecipePolicy with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:containerRecipeArn:GetContainerRecipePolicy', getContainerRecipePolicy_containerRecipeArn - The Amazon Resource Name (ARN) of the container recipe for the policy being requested.

Request Lenses

getContainerRecipePolicy_containerRecipeArn :: Lens' GetContainerRecipePolicy Text

The Amazon Resource Name (ARN) of the container recipe for the policy being requested.

Destructuring the Response

data GetContainerRecipePolicyResponse

See: newGetContainerRecipePolicyResponse smart constructor.

Constructors

GetContainerRecipePolicyResponse'

Fields

requestId :: Maybe Text

The request ID that uniquely identifies this request.

policy :: Maybe Text

The container recipe policy object that is returned.

httpStatus :: Int

The response's http status code.

newGetContainerRecipePolicyResponse

Arguments

:: Int
$sel:httpStatus:GetContainerRecipePolicyResponse'

-> GetContainerRecipePolicyResponse

Create a value of GetContainerRecipePolicyResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:requestId:GetContainerRecipePolicyResponse', getContainerRecipePolicyResponse_requestId - The request ID that uniquely identifies this request.

$sel:policy:GetContainerRecipePolicyResponse', getContainerRecipePolicyResponse_policy - The container recipe policy object that is returned.

$sel:httpStatus:GetContainerRecipePolicyResponse', getContainerRecipePolicyResponse_httpStatus - The response's http status code.

Response Lenses

getContainerRecipePolicyResponse_requestId :: Lens' GetContainerRecipePolicyResponse (Maybe Text)

The request ID that uniquely identifies this request.

getContainerRecipePolicyResponse_policy :: Lens' GetContainerRecipePolicyResponse (Maybe Text)

The container recipe policy object that is returned.

getContainerRecipePolicyResponse_httpStatus :: Lens' GetContainerRecipePolicyResponse Int

The response's http status code.