Amazonka.ImageBuilder.ListComponentBuildVersions

Description

Returns the list of component build versions for the specified semantic version.

The semantic version has four nodes: <major>.<minor>.<patch>/<build>. You can assign values for the first three, and can filter on all of them.

Filtering: With semantic versioning, you have the flexibility to use wildcards (x) to specify the most recent versions or nodes when selecting the base image or components for your recipe. When you use a wildcard in any node, all nodes to the right of the first wildcard must also be wildcards.

Creating a Request

data ListComponentBuildVersions

See: newListComponentBuildVersions smart constructor.

Constructors

ListComponentBuildVersions'

Fields

nextToken :: Maybe Text

A token to specify where to start paginating. This is the NextToken from a previously truncated response.

maxResults :: Maybe Natural

The maximum items to return in a request.

componentVersionArn :: Text

The component version Amazon Resource Name (ARN) whose versions you want to list.

newListComponentBuildVersions

Arguments

:: Text
$sel:componentVersionArn:ListComponentBuildVersions'

-> ListComponentBuildVersions

Create a value of ListComponentBuildVersions with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:nextToken:ListComponentBuildVersions', listComponentBuildVersions_nextToken - A token to specify where to start paginating. This is the NextToken from a previously truncated response.

$sel:maxResults:ListComponentBuildVersions', listComponentBuildVersions_maxResults - The maximum items to return in a request.

$sel:componentVersionArn:ListComponentBuildVersions', listComponentBuildVersions_componentVersionArn - The component version Amazon Resource Name (ARN) whose versions you want to list.

Request Lenses

listComponentBuildVersions_nextToken :: Lens' ListComponentBuildVersions (Maybe Text)

A token to specify where to start paginating. This is the NextToken from a previously truncated response.

listComponentBuildVersions_maxResults :: Lens' ListComponentBuildVersions (Maybe Natural)

The maximum items to return in a request.

listComponentBuildVersions_componentVersionArn :: Lens' ListComponentBuildVersions Text

The component version Amazon Resource Name (ARN) whose versions you want to list.

Destructuring the Response

data ListComponentBuildVersionsResponse

See: newListComponentBuildVersionsResponse smart constructor.

Constructors

ListComponentBuildVersionsResponse'

Fields

requestId :: Maybe Text

The request ID that uniquely identifies this request.

nextToken :: Maybe Text

The next token used for paginated responses. When this is not empty, there are additional elements that the service has not included in this request. Use this token with the next request to retrieve additional objects.

componentSummaryList :: Maybe [ComponentSummary]

The list of component summaries for the specified semantic version.

httpStatus :: Int

The response's http status code.

newListComponentBuildVersionsResponse

Arguments

:: Int
$sel:httpStatus:ListComponentBuildVersionsResponse'

-> ListComponentBuildVersionsResponse

Create a value of ListComponentBuildVersionsResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:requestId:ListComponentBuildVersionsResponse', listComponentBuildVersionsResponse_requestId - The request ID that uniquely identifies this request.

$sel:nextToken:ListComponentBuildVersions', listComponentBuildVersionsResponse_nextToken - The next token used for paginated responses. When this is not empty, there are additional elements that the service has not included in this request. Use this token with the next request to retrieve additional objects.

$sel:componentSummaryList:ListComponentBuildVersionsResponse', listComponentBuildVersionsResponse_componentSummaryList - The list of component summaries for the specified semantic version.

$sel:httpStatus:ListComponentBuildVersionsResponse', listComponentBuildVersionsResponse_httpStatus - The response's http status code.

Response Lenses

listComponentBuildVersionsResponse_requestId :: Lens' ListComponentBuildVersionsResponse (Maybe Text)

The request ID that uniquely identifies this request.

listComponentBuildVersionsResponse_nextToken :: Lens' ListComponentBuildVersionsResponse (Maybe Text)

The next token used for paginated responses. When this is not empty, there are additional elements that the service has not included in this request. Use this token with the next request to retrieve additional objects.

listComponentBuildVersionsResponse_componentSummaryList :: Lens' ListComponentBuildVersionsResponse (Maybe [ComponentSummary])

The list of component summaries for the specified semantic version.

listComponentBuildVersionsResponse_httpStatus :: Lens' ListComponentBuildVersionsResponse Int

The response's http status code.