Amazonka.ImageBuilder.ListImages

Description

Returns the list of images that you have access to.

Creating a Request

data ListImages

See: newListImages smart constructor.

Constructors

ListImages'

Fields

filters :: Maybe (NonEmpty Filter)

Use the following filters to streamline results:

name

osVersion

platform

type

version

owner :: Maybe Ownership

The owner defines which images you want to list. By default, this request will only show images owned by your account. You can use this field to specify if you want to view images owned by yourself, by Amazon, or those images that have been shared with you by other customers.

byName :: Maybe Bool

Requests a list of images with a specific recipe name.

nextToken :: Maybe Text

A token to specify where to start paginating. This is the NextToken from a previously truncated response.

includeDeprecated :: Maybe Bool

Includes deprecated images in the response list.

maxResults :: Maybe Natural

The maximum items to return in a request.

newListImages :: ListImages

Create a value of ListImages with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:filters:ListImages', listImages_filters - Use the following filters to streamline results:

name

osVersion

platform

type

version

$sel:owner:ListImages', listImages_owner - The owner defines which images you want to list. By default, this request will only show images owned by your account. You can use this field to specify if you want to view images owned by yourself, by Amazon, or those images that have been shared with you by other customers.

$sel:byName:ListImages', listImages_byName - Requests a list of images with a specific recipe name.

$sel:nextToken:ListImages', listImages_nextToken - A token to specify where to start paginating. This is the NextToken from a previously truncated response.

$sel:includeDeprecated:ListImages', listImages_includeDeprecated - Includes deprecated images in the response list.

$sel:maxResults:ListImages', listImages_maxResults - The maximum items to return in a request.

Request Lenses

listImages_filters :: Lens' ListImages (Maybe (NonEmpty Filter))

Use the following filters to streamline results:

name

osVersion

platform

type

version

listImages_owner :: Lens' ListImages (Maybe Ownership)

The owner defines which images you want to list. By default, this request will only show images owned by your account. You can use this field to specify if you want to view images owned by yourself, by Amazon, or those images that have been shared with you by other customers.

listImages_byName :: Lens' ListImages (Maybe Bool)

Requests a list of images with a specific recipe name.

listImages_nextToken :: Lens' ListImages (Maybe Text)

A token to specify where to start paginating. This is the NextToken from a previously truncated response.

listImages_includeDeprecated :: Lens' ListImages (Maybe Bool)

Includes deprecated images in the response list.

listImages_maxResults :: Lens' ListImages (Maybe Natural)

The maximum items to return in a request.

Destructuring the Response

data ListImagesResponse

See: newListImagesResponse smart constructor.

Constructors

ListImagesResponse'

Fields

requestId :: Maybe Text

The request ID that uniquely identifies this request.

imageVersionList :: Maybe [ImageVersion]

The list of image semantic versions.

The semantic version has four nodes: <major>.<minor>.<patch>/<build>. You can assign values for the first three, and can filter on all of them.

Filtering: With semantic versioning, you have the flexibility to use wildcards (x) to specify the most recent versions or nodes when selecting the base image or components for your recipe. When you use a wildcard in any node, all nodes to the right of the first wildcard must also be wildcards.

nextToken :: Maybe Text

The next token used for paginated responses. When this is not empty, there are additional elements that the service has not included in this request. Use this token with the next request to retrieve additional objects.

httpStatus :: Int

The response's http status code.

newListImagesResponse

Arguments

:: Int
$sel:httpStatus:ListImagesResponse'

-> ListImagesResponse

Create a value of ListImagesResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:requestId:ListImagesResponse', listImagesResponse_requestId - The request ID that uniquely identifies this request.

$sel:imageVersionList:ListImagesResponse', listImagesResponse_imageVersionList - The list of image semantic versions.

The semantic version has four nodes: <major>.<minor>.<patch>/<build>. You can assign values for the first three, and can filter on all of them.

Filtering: With semantic versioning, you have the flexibility to use wildcards (x) to specify the most recent versions or nodes when selecting the base image or components for your recipe. When you use a wildcard in any node, all nodes to the right of the first wildcard must also be wildcards.

$sel:nextToken:ListImages', listImagesResponse_nextToken - The next token used for paginated responses. When this is not empty, there are additional elements that the service has not included in this request. Use this token with the next request to retrieve additional objects.

$sel:httpStatus:ListImagesResponse', listImagesResponse_httpStatus - The response's http status code.

Response Lenses

listImagesResponse_requestId :: Lens' ListImagesResponse (Maybe Text)

The request ID that uniquely identifies this request.

listImagesResponse_imageVersionList :: Lens' ListImagesResponse (Maybe [ImageVersion])

The list of image semantic versions.

The semantic version has four nodes: <major>.<minor>.<patch>/<build>. You can assign values for the first three, and can filter on all of them.

Filtering: With semantic versioning, you have the flexibility to use wildcards (x) to specify the most recent versions or nodes when selecting the base image or components for your recipe. When you use a wildcard in any node, all nodes to the right of the first wildcard must also be wildcards.

listImagesResponse_nextToken :: Lens' ListImagesResponse (Maybe Text)

The next token used for paginated responses. When this is not empty, there are additional elements that the service has not included in this request. Use this token with the next request to retrieve additional objects.

listImagesResponse_httpStatus :: Lens' ListImagesResponse Int

The response's http status code.