Amazonka.ImageBuilder.PutComponentPolicy

Description

Applies a policy to a component. We recommend that you call the RAM API CreateResourceShare to share resources. If you call the Image Builder API PutComponentPolicy, you must also call the RAM API PromoteResourceShareCreatedFromPolicy in order for the resource to be visible to all principals with whom the resource is shared.

Creating a Request

data PutComponentPolicy

See: newPutComponentPolicy smart constructor.

Constructors

PutComponentPolicy'

Fields

componentArn :: Text

The Amazon Resource Name (ARN) of the component that this policy should be applied to.

policy :: Text

The policy to apply.

newPutComponentPolicy

Arguments

:: Text
$sel:componentArn:PutComponentPolicy'

-> Text
$sel:policy:PutComponentPolicy'

-> PutComponentPolicy

Create a value of PutComponentPolicy with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:componentArn:PutComponentPolicy', putComponentPolicy_componentArn - The Amazon Resource Name (ARN) of the component that this policy should be applied to.

$sel:policy:PutComponentPolicy', putComponentPolicy_policy - The policy to apply.

Request Lenses

putComponentPolicy_componentArn :: Lens' PutComponentPolicy Text

The Amazon Resource Name (ARN) of the component that this policy should be applied to.

putComponentPolicy_policy :: Lens' PutComponentPolicy Text

The policy to apply.

Destructuring the Response

data PutComponentPolicyResponse

See: newPutComponentPolicyResponse smart constructor.

Constructors

PutComponentPolicyResponse'

Fields

requestId :: Maybe Text

The request ID that uniquely identifies this request.

componentArn :: Maybe Text

The Amazon Resource Name (ARN) of the component that this policy was applied to.

httpStatus :: Int

The response's http status code.

newPutComponentPolicyResponse

Arguments

:: Int
$sel:httpStatus:PutComponentPolicyResponse'

-> PutComponentPolicyResponse

Create a value of PutComponentPolicyResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:requestId:PutComponentPolicyResponse', putComponentPolicyResponse_requestId - The request ID that uniquely identifies this request.

$sel:componentArn:PutComponentPolicy', putComponentPolicyResponse_componentArn - The Amazon Resource Name (ARN) of the component that this policy was applied to.

$sel:httpStatus:PutComponentPolicyResponse', putComponentPolicyResponse_httpStatus - The response's http status code.

Response Lenses

putComponentPolicyResponse_requestId :: Lens' PutComponentPolicyResponse (Maybe Text)

The request ID that uniquely identifies this request.

putComponentPolicyResponse_componentArn :: Lens' PutComponentPolicyResponse (Maybe Text)

The Amazon Resource Name (ARN) of the component that this policy was applied to.

putComponentPolicyResponse_httpStatus :: Lens' PutComponentPolicyResponse Int

The response's http status code.