Amazonka.ImageBuilder.Types.AmiDistributionConfiguration

Description

Documentation

data AmiDistributionConfiguration

Define and configure the output AMIs of the pipeline.

See: newAmiDistributionConfiguration smart constructor.

Constructors

AmiDistributionConfiguration'

Fields

launchPermission :: Maybe LaunchPermissionConfiguration

Launch permissions can be used to configure which Amazon Web Services accounts can use the AMI to launch instances.

targetAccountIds :: Maybe (NonEmpty Text)

The ID of an account to which you want to distribute an image.

amiTags :: Maybe (HashMap Text Text)

The tags to apply to AMIs distributed to this Region.

name :: Maybe Text

The name of the output AMI.

kmsKeyId :: Maybe Text

The KMS key identifier used to encrypt the distributed image.

description :: Maybe Text

The description of the distribution configuration. Minimum and maximum length are in characters.

newAmiDistributionConfiguration :: AmiDistributionConfiguration

Create a value of AmiDistributionConfiguration with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:launchPermission:AmiDistributionConfiguration', amiDistributionConfiguration_launchPermission - Launch permissions can be used to configure which Amazon Web Services accounts can use the AMI to launch instances.

$sel:targetAccountIds:AmiDistributionConfiguration', amiDistributionConfiguration_targetAccountIds - The ID of an account to which you want to distribute an image.

$sel:amiTags:AmiDistributionConfiguration', amiDistributionConfiguration_amiTags - The tags to apply to AMIs distributed to this Region.

$sel:name:AmiDistributionConfiguration', amiDistributionConfiguration_name - The name of the output AMI.

$sel:kmsKeyId:AmiDistributionConfiguration', amiDistributionConfiguration_kmsKeyId - The KMS key identifier used to encrypt the distributed image.

$sel:description:AmiDistributionConfiguration', amiDistributionConfiguration_description - The description of the distribution configuration. Minimum and maximum length are in characters.

amiDistributionConfiguration_launchPermission :: Lens' AmiDistributionConfiguration (Maybe LaunchPermissionConfiguration)

Launch permissions can be used to configure which Amazon Web Services accounts can use the AMI to launch instances.

amiDistributionConfiguration_targetAccountIds :: Lens' AmiDistributionConfiguration (Maybe (NonEmpty Text))

The ID of an account to which you want to distribute an image.

amiDistributionConfiguration_amiTags :: Lens' AmiDistributionConfiguration (Maybe (HashMap Text Text))

The tags to apply to AMIs distributed to this Region.

amiDistributionConfiguration_name :: Lens' AmiDistributionConfiguration (Maybe Text)

The name of the output AMI.

amiDistributionConfiguration_kmsKeyId :: Lens' AmiDistributionConfiguration (Maybe Text)

The KMS key identifier used to encrypt the distributed image.

amiDistributionConfiguration_description :: Lens' AmiDistributionConfiguration (Maybe Text)

The description of the distribution configuration. Minimum and maximum length are in characters.