Amazonka.ImageBuilder.Types.Component

Description

Documentation

data Component

A detailed view of a component.

See: newComponent smart constructor.

Constructors

Component'

Fields

state :: Maybe ComponentState

Describes the current status of the component. This is used for components that are no longer active.

platform :: Maybe Platform

The platform of the component.

arn :: Maybe Text

The Amazon Resource Name (ARN) of the component.

data' :: Maybe Text

The data of the component.

supportedOsVersions :: Maybe (NonEmpty Text)

The operating system (OS) version supported by the component. If the OS information is available, a prefix match is performed against the base image OS version during image recipe creation.

encrypted :: Maybe Bool

The encryption status of the component.

owner :: Maybe Text

The owner of the component.

dateCreated :: Maybe Text

The date that the component was created.

name :: Maybe Text

The name of the component.

kmsKeyId :: Maybe Text

The KMS key identifier used to encrypt the component.

version :: Maybe Text

The version of the component.

parameters :: Maybe [ComponentParameterDetail]

Contains parameter details for each of the parameters that are defined for the component.

changeDescription :: Maybe Text

The change description of the component.

type' :: Maybe ComponentType

The type of the component denotes whether the component is used to build the image or only to test it.

description :: Maybe Text

The description of the component.

tags :: Maybe (HashMap Text Text)

The tags associated with the component.

newComponent :: Component

Create a value of Component with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:state:Component', component_state - Describes the current status of the component. This is used for components that are no longer active.

$sel:platform:Component', component_platform - The platform of the component.

$sel:arn:Component', component_arn - The Amazon Resource Name (ARN) of the component.

$sel:data':Component', component_data - The data of the component.

$sel:supportedOsVersions:Component', component_supportedOsVersions - The operating system (OS) version supported by the component. If the OS information is available, a prefix match is performed against the base image OS version during image recipe creation.

$sel:encrypted:Component', component_encrypted - The encryption status of the component.

$sel:owner:Component', component_owner - The owner of the component.

$sel:dateCreated:Component', component_dateCreated - The date that the component was created.

$sel:name:Component', component_name - The name of the component.

$sel:kmsKeyId:Component', component_kmsKeyId - The KMS key identifier used to encrypt the component.

$sel:version:Component', component_version - The version of the component.

$sel:parameters:Component', component_parameters - Contains parameter details for each of the parameters that are defined for the component.

$sel:changeDescription:Component', component_changeDescription - The change description of the component.

$sel:type':Component', component_type - The type of the component denotes whether the component is used to build the image or only to test it.

$sel:description:Component', component_description - The description of the component.

$sel:tags:Component', component_tags - The tags associated with the component.

component_state :: Lens' Component (Maybe ComponentState)

Describes the current status of the component. This is used for components that are no longer active.

component_platform :: Lens' Component (Maybe Platform)

The platform of the component.

component_arn :: Lens' Component (Maybe Text)

The Amazon Resource Name (ARN) of the component.

component_data :: Lens' Component (Maybe Text)

The data of the component.

component_supportedOsVersions :: Lens' Component (Maybe (NonEmpty Text))

The operating system (OS) version supported by the component. If the OS information is available, a prefix match is performed against the base image OS version during image recipe creation.

component_encrypted :: Lens' Component (Maybe Bool)

The encryption status of the component.

component_owner :: Lens' Component (Maybe Text)

The owner of the component.

component_dateCreated :: Lens' Component (Maybe Text)

The date that the component was created.

component_name :: Lens' Component (Maybe Text)

The name of the component.

component_kmsKeyId :: Lens' Component (Maybe Text)

The KMS key identifier used to encrypt the component.

component_version :: Lens' Component (Maybe Text)

The version of the component.

component_parameters :: Lens' Component (Maybe [ComponentParameterDetail])

Contains parameter details for each of the parameters that are defined for the component.

component_changeDescription :: Lens' Component (Maybe Text)

The change description of the component.

component_type :: Lens' Component (Maybe ComponentType)

The type of the component denotes whether the component is used to build the image or only to test it.

component_description :: Lens' Component (Maybe Text)

The description of the component.

component_tags :: Lens' Component (Maybe (HashMap Text Text))

The tags associated with the component.