Amazonka.ImageBuilder.Types.ComponentConfiguration

Description

Documentation

data ComponentConfiguration

Configuration details of the component.

See: newComponentConfiguration smart constructor.

Constructors

ComponentConfiguration'

Fields

parameters :: Maybe (NonEmpty ComponentParameter)

A group of parameter settings that are used to configure the component for a specific recipe.

componentArn :: Text

The Amazon Resource Name (ARN) of the component.

newComponentConfiguration

Arguments

:: Text
$sel:componentArn:ComponentConfiguration'

-> ComponentConfiguration

Create a value of ComponentConfiguration with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:parameters:ComponentConfiguration', componentConfiguration_parameters - A group of parameter settings that are used to configure the component for a specific recipe.

$sel:componentArn:ComponentConfiguration', componentConfiguration_componentArn - The Amazon Resource Name (ARN) of the component.

componentConfiguration_parameters :: Lens' ComponentConfiguration (Maybe (NonEmpty ComponentParameter))

A group of parameter settings that are used to configure the component for a specific recipe.

componentConfiguration_componentArn :: Lens' ComponentConfiguration Text

The Amazon Resource Name (ARN) of the component.