Amazonka.ImageBuilder.Types.ComponentVersion

Description

Documentation

data ComponentVersion

The defining characteristics of a specific version of an Amazon Web Services TOE component.

See: newComponentVersion smart constructor.

Constructors

ComponentVersion'

Fields

platform :: Maybe Platform

The platform of the component.

arn :: Maybe Text

The Amazon Resource Name (ARN) of the component.

Semantic versioning is included in each object's Amazon Resource Name (ARN), at the level that applies to that object as follows:

Versionless ARNs and Name ARNs do not include specific values in any of the nodes. The nodes are either left off entirely, or they are specified as wildcards, for example: x.x.x.

Version ARNs have only the first three nodes: <major>.<minor>.<patch>

Build version ARNs have all four nodes, and point to a specific build for a specific version of an object.

supportedOsVersions :: Maybe (NonEmpty Text)

he operating system (OS) version supported by the component. If the OS information is available, a prefix match is performed against the base image OS version during image recipe creation.

owner :: Maybe Text

The owner of the component.

dateCreated :: Maybe Text

The date that the component was created.

name :: Maybe Text

The name of the component.

version :: Maybe Text

The semantic version of the component.

The semantic version has four nodes: <major>.<minor>.<patch>/<build>. You can assign values for the first three, and can filter on all of them.

Assignment: For the first three nodes you can assign any positive integer value, including zero, with an upper limit of 2^30-1, or 1073741823 for each node. Image Builder automatically assigns the build number to the fourth node.

Patterns: You can use any numeric pattern that adheres to the assignment requirements for the nodes that you can assign. For example, you might choose a software version pattern, such as 1.0.0, or a date, such as 2021.01.01.

Filtering: With semantic versioning, you have the flexibility to use wildcards (x) to specify the most recent versions or nodes when selecting the base image or components for your recipe. When you use a wildcard in any node, all nodes to the right of the first wildcard must also be wildcards.

type' :: Maybe ComponentType

The type of the component denotes whether the component is used to build the image or only to test it.

description :: Maybe Text

The description of the component.

newComponentVersion :: ComponentVersion

Create a value of ComponentVersion with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:platform:ComponentVersion', componentVersion_platform - The platform of the component.

$sel:arn:ComponentVersion', componentVersion_arn - The Amazon Resource Name (ARN) of the component.

Semantic versioning is included in each object's Amazon Resource Name (ARN), at the level that applies to that object as follows:

Versionless ARNs and Name ARNs do not include specific values in any of the nodes. The nodes are either left off entirely, or they are specified as wildcards, for example: x.x.x.

Version ARNs have only the first three nodes: <major>.<minor>.<patch>

Build version ARNs have all four nodes, and point to a specific build for a specific version of an object.

$sel:supportedOsVersions:ComponentVersion', componentVersion_supportedOsVersions - he operating system (OS) version supported by the component. If the OS information is available, a prefix match is performed against the base image OS version during image recipe creation.

$sel:owner:ComponentVersion', componentVersion_owner - The owner of the component.

$sel:dateCreated:ComponentVersion', componentVersion_dateCreated - The date that the component was created.

$sel:name:ComponentVersion', componentVersion_name - The name of the component.

$sel:version:ComponentVersion', componentVersion_version - The semantic version of the component.

The semantic version has four nodes: <major>.<minor>.<patch>/<build>. You can assign values for the first three, and can filter on all of them.

Assignment: For the first three nodes you can assign any positive integer value, including zero, with an upper limit of 2^30-1, or 1073741823 for each node. Image Builder automatically assigns the build number to the fourth node.

Patterns: You can use any numeric pattern that adheres to the assignment requirements for the nodes that you can assign. For example, you might choose a software version pattern, such as 1.0.0, or a date, such as 2021.01.01.

Filtering: With semantic versioning, you have the flexibility to use wildcards (x) to specify the most recent versions or nodes when selecting the base image or components for your recipe. When you use a wildcard in any node, all nodes to the right of the first wildcard must also be wildcards.

$sel:type':ComponentVersion', componentVersion_type - The type of the component denotes whether the component is used to build the image or only to test it.

$sel:description:ComponentVersion', componentVersion_description - The description of the component.

componentVersion_platform :: Lens' ComponentVersion (Maybe Platform)

The platform of the component.

componentVersion_arn :: Lens' ComponentVersion (Maybe Text)

The Amazon Resource Name (ARN) of the component.

Semantic versioning is included in each object's Amazon Resource Name (ARN), at the level that applies to that object as follows:

Versionless ARNs and Name ARNs do not include specific values in any of the nodes. The nodes are either left off entirely, or they are specified as wildcards, for example: x.x.x.

Version ARNs have only the first three nodes: <major>.<minor>.<patch>

Build version ARNs have all four nodes, and point to a specific build for a specific version of an object.

componentVersion_supportedOsVersions :: Lens' ComponentVersion (Maybe (NonEmpty Text))

he operating system (OS) version supported by the component. If the OS information is available, a prefix match is performed against the base image OS version during image recipe creation.

componentVersion_owner :: Lens' ComponentVersion (Maybe Text)

The owner of the component.

componentVersion_dateCreated :: Lens' ComponentVersion (Maybe Text)

The date that the component was created.

componentVersion_name :: Lens' ComponentVersion (Maybe Text)

The name of the component.

componentVersion_version :: Lens' ComponentVersion (Maybe Text)

The semantic version of the component.

The semantic version has four nodes: <major>.<minor>.<patch>/<build>. You can assign values for the first three, and can filter on all of them.

Assignment: For the first three nodes you can assign any positive integer value, including zero, with an upper limit of 2^30-1, or 1073741823 for each node. Image Builder automatically assigns the build number to the fourth node.

Patterns: You can use any numeric pattern that adheres to the assignment requirements for the nodes that you can assign. For example, you might choose a software version pattern, such as 1.0.0, or a date, such as 2021.01.01.

Filtering: With semantic versioning, you have the flexibility to use wildcards (x) to specify the most recent versions or nodes when selecting the base image or components for your recipe. When you use a wildcard in any node, all nodes to the right of the first wildcard must also be wildcards.

componentVersion_type :: Lens' ComponentVersion (Maybe ComponentType)

The type of the component denotes whether the component is used to build the image or only to test it.

componentVersion_description :: Lens' ComponentVersion (Maybe Text)

The description of the component.