Amazonka.ImageBuilder.Types.Container

Description

Documentation

data Container

A container encapsulates the runtime environment for an application.

See: newContainer smart constructor.

Constructors

Container'

Fields

imageUris :: Maybe [Text]

A list of URIs for containers created in the context Region.

region :: Maybe Text

Containers and container images are Region-specific. This is the Region context for the container.

newContainer :: Container

Create a value of Container with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:imageUris:Container', container_imageUris - A list of URIs for containers created in the context Region.

$sel:region:Container', container_region - Containers and container images are Region-specific. This is the Region context for the container.

container_imageUris :: Lens' Container (Maybe [Text])

A list of URIs for containers created in the context Region.

container_region :: Lens' Container (Maybe Text)

Containers and container images are Region-specific. This is the Region context for the container.