Amazonka.ImageBuilder.Types.EbsInstanceBlockDeviceSpecification

Description

Documentation

data EbsInstanceBlockDeviceSpecification

Amazon EBS-specific block device mapping specifications.

See: newEbsInstanceBlockDeviceSpecification smart constructor.

Constructors

EbsInstanceBlockDeviceSpecification'

Fields

deleteOnTermination :: Maybe Bool

Use to configure delete on termination of the associated device.

throughput :: Maybe Natural

For GP3 volumes only – The throughput in MiB/s that the volume supports.

volumeSize :: Maybe Natural

Use to override the device's volume size.

iops :: Maybe Natural

Use to configure device IOPS.

encrypted :: Maybe Bool

Use to configure device encryption.

kmsKeyId :: Maybe Text

Use to configure the KMS key to use when encrypting the device.

volumeType :: Maybe EbsVolumeType

Use to override the device's volume type.

snapshotId :: Maybe Text

The snapshot that defines the device contents.

newEbsInstanceBlockDeviceSpecification :: EbsInstanceBlockDeviceSpecification

Create a value of EbsInstanceBlockDeviceSpecification with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:deleteOnTermination:EbsInstanceBlockDeviceSpecification', ebsInstanceBlockDeviceSpecification_deleteOnTermination - Use to configure delete on termination of the associated device.

$sel:throughput:EbsInstanceBlockDeviceSpecification', ebsInstanceBlockDeviceSpecification_throughput - For GP3 volumes only – The throughput in MiB/s that the volume supports.

$sel:volumeSize:EbsInstanceBlockDeviceSpecification', ebsInstanceBlockDeviceSpecification_volumeSize - Use to override the device's volume size.

$sel:iops:EbsInstanceBlockDeviceSpecification', ebsInstanceBlockDeviceSpecification_iops - Use to configure device IOPS.

$sel:encrypted:EbsInstanceBlockDeviceSpecification', ebsInstanceBlockDeviceSpecification_encrypted - Use to configure device encryption.

$sel:kmsKeyId:EbsInstanceBlockDeviceSpecification', ebsInstanceBlockDeviceSpecification_kmsKeyId - Use to configure the KMS key to use when encrypting the device.

$sel:volumeType:EbsInstanceBlockDeviceSpecification', ebsInstanceBlockDeviceSpecification_volumeType - Use to override the device's volume type.

$sel:snapshotId:EbsInstanceBlockDeviceSpecification', ebsInstanceBlockDeviceSpecification_snapshotId - The snapshot that defines the device contents.

ebsInstanceBlockDeviceSpecification_deleteOnTermination :: Lens' EbsInstanceBlockDeviceSpecification (Maybe Bool)

Use to configure delete on termination of the associated device.

ebsInstanceBlockDeviceSpecification_throughput :: Lens' EbsInstanceBlockDeviceSpecification (Maybe Natural)

For GP3 volumes only – The throughput in MiB/s that the volume supports.

ebsInstanceBlockDeviceSpecification_volumeSize :: Lens' EbsInstanceBlockDeviceSpecification (Maybe Natural)

Use to override the device's volume size.

ebsInstanceBlockDeviceSpecification_iops :: Lens' EbsInstanceBlockDeviceSpecification (Maybe Natural)

Use to configure device IOPS.

ebsInstanceBlockDeviceSpecification_encrypted :: Lens' EbsInstanceBlockDeviceSpecification (Maybe Bool)

Use to configure device encryption.

ebsInstanceBlockDeviceSpecification_kmsKeyId :: Lens' EbsInstanceBlockDeviceSpecification (Maybe Text)

Use to configure the KMS key to use when encrypting the device.

ebsInstanceBlockDeviceSpecification_volumeType :: Lens' EbsInstanceBlockDeviceSpecification (Maybe EbsVolumeType)

Use to override the device's volume type.

ebsInstanceBlockDeviceSpecification_snapshotId :: Lens' EbsInstanceBlockDeviceSpecification (Maybe Text)

The snapshot that defines the device contents.