Amazonka.ImageBuilder.Types.ImageRecipeSummary

Description

Documentation

data ImageRecipeSummary

A summary of an image recipe.

See: newImageRecipeSummary smart constructor.

Constructors

ImageRecipeSummary'

Fields

platform :: Maybe Platform

The platform of the image recipe.

arn :: Maybe Text

The Amazon Resource Name (ARN) of the image recipe.

parentImage :: Maybe Text

The base image of the image recipe.

owner :: Maybe Text

The owner of the image recipe.

dateCreated :: Maybe Text

The date on which this image recipe was created.

name :: Maybe Text

The name of the image recipe.

tags :: Maybe (HashMap Text Text)

The tags of the image recipe.

newImageRecipeSummary :: ImageRecipeSummary

Create a value of ImageRecipeSummary with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:platform:ImageRecipeSummary', imageRecipeSummary_platform - The platform of the image recipe.

$sel:arn:ImageRecipeSummary', imageRecipeSummary_arn - The Amazon Resource Name (ARN) of the image recipe.

$sel:parentImage:ImageRecipeSummary', imageRecipeSummary_parentImage - The base image of the image recipe.

$sel:owner:ImageRecipeSummary', imageRecipeSummary_owner - The owner of the image recipe.

$sel:dateCreated:ImageRecipeSummary', imageRecipeSummary_dateCreated - The date on which this image recipe was created.

$sel:name:ImageRecipeSummary', imageRecipeSummary_name - The name of the image recipe.

$sel:tags:ImageRecipeSummary', imageRecipeSummary_tags - The tags of the image recipe.

imageRecipeSummary_platform :: Lens' ImageRecipeSummary (Maybe Platform)

The platform of the image recipe.

imageRecipeSummary_arn :: Lens' ImageRecipeSummary (Maybe Text)

The Amazon Resource Name (ARN) of the image recipe.

imageRecipeSummary_parentImage :: Lens' ImageRecipeSummary (Maybe Text)

The base image of the image recipe.

imageRecipeSummary_owner :: Lens' ImageRecipeSummary (Maybe Text)

The owner of the image recipe.

imageRecipeSummary_dateCreated :: Lens' ImageRecipeSummary (Maybe Text)

The date on which this image recipe was created.

imageRecipeSummary_name :: Lens' ImageRecipeSummary (Maybe Text)

The name of the image recipe.

imageRecipeSummary_tags :: Lens' ImageRecipeSummary (Maybe (HashMap Text Text))

The tags of the image recipe.