Amazonka.ImageBuilder.Types.ImageVersion

Description

Documentation

data ImageVersion

The defining characteristics of a specific version of an Image Builder image.

See: newImageVersion smart constructor.

Constructors

ImageVersion'

Fields

platform :: Maybe Platform

The platform of the image version, for example "Windows" or "Linux".

arn :: Maybe Text

The Amazon Resource Name (ARN) of a specific version of an Image Builder image.

Semantic versioning is included in each object's Amazon Resource Name (ARN), at the level that applies to that object as follows:

Versionless ARNs and Name ARNs do not include specific values in any of the nodes. The nodes are either left off entirely, or they are specified as wildcards, for example: x.x.x.

Version ARNs have only the first three nodes: <major>.<minor>.<patch>

Build version ARNs have all four nodes, and point to a specific build for a specific version of an object.

owner :: Maybe Text

The owner of the image version.

dateCreated :: Maybe Text

The date on which this specific version of the Image Builder image was created.

name :: Maybe Text

The name of this specific version of an Image Builder image.

version :: Maybe Text

Details for a specific version of an Image Builder image. This version follows the semantic version syntax.

The semantic version has four nodes: <major>.<minor>.<patch>/<build>. You can assign values for the first three, and can filter on all of them.

Assignment: For the first three nodes you can assign any positive integer value, including zero, with an upper limit of 2^30-1, or 1073741823 for each node. Image Builder automatically assigns the build number to the fourth node.

Patterns: You can use any numeric pattern that adheres to the assignment requirements for the nodes that you can assign. For example, you might choose a software version pattern, such as 1.0.0, or a date, such as 2021.01.01.

Filtering: With semantic versioning, you have the flexibility to use wildcards (x) to specify the most recent versions or nodes when selecting the base image or components for your recipe. When you use a wildcard in any node, all nodes to the right of the first wildcard must also be wildcards.

type' :: Maybe ImageType

Specifies whether this image is an AMI or a container image.

osVersion :: Maybe Text

The operating system version of the Amazon EC2 build instance. For example, Amazon Linux 2, Ubuntu 18, or Microsoft Windows Server 2019.

newImageVersion :: ImageVersion

Create a value of ImageVersion with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:platform:ImageVersion', imageVersion_platform - The platform of the image version, for example "Windows" or "Linux".

$sel:arn:ImageVersion', imageVersion_arn - The Amazon Resource Name (ARN) of a specific version of an Image Builder image.

Semantic versioning is included in each object's Amazon Resource Name (ARN), at the level that applies to that object as follows:

Versionless ARNs and Name ARNs do not include specific values in any of the nodes. The nodes are either left off entirely, or they are specified as wildcards, for example: x.x.x.

Version ARNs have only the first three nodes: <major>.<minor>.<patch>

Build version ARNs have all four nodes, and point to a specific build for a specific version of an object.

$sel:owner:ImageVersion', imageVersion_owner - The owner of the image version.

$sel:dateCreated:ImageVersion', imageVersion_dateCreated - The date on which this specific version of the Image Builder image was created.

$sel:name:ImageVersion', imageVersion_name - The name of this specific version of an Image Builder image.

$sel:version:ImageVersion', imageVersion_version - Details for a specific version of an Image Builder image. This version follows the semantic version syntax.

The semantic version has four nodes: <major>.<minor>.<patch>/<build>. You can assign values for the first three, and can filter on all of them.

Assignment: For the first three nodes you can assign any positive integer value, including zero, with an upper limit of 2^30-1, or 1073741823 for each node. Image Builder automatically assigns the build number to the fourth node.

Patterns: You can use any numeric pattern that adheres to the assignment requirements for the nodes that you can assign. For example, you might choose a software version pattern, such as 1.0.0, or a date, such as 2021.01.01.

Filtering: With semantic versioning, you have the flexibility to use wildcards (x) to specify the most recent versions or nodes when selecting the base image or components for your recipe. When you use a wildcard in any node, all nodes to the right of the first wildcard must also be wildcards.

$sel:type':ImageVersion', imageVersion_type - Specifies whether this image is an AMI or a container image.

$sel:osVersion:ImageVersion', imageVersion_osVersion - The operating system version of the Amazon EC2 build instance. For example, Amazon Linux 2, Ubuntu 18, or Microsoft Windows Server 2019.

imageVersion_platform :: Lens' ImageVersion (Maybe Platform)

The platform of the image version, for example "Windows" or "Linux".

imageVersion_arn :: Lens' ImageVersion (Maybe Text)

The Amazon Resource Name (ARN) of a specific version of an Image Builder image.

Semantic versioning is included in each object's Amazon Resource Name (ARN), at the level that applies to that object as follows:

Versionless ARNs and Name ARNs do not include specific values in any of the nodes. The nodes are either left off entirely, or they are specified as wildcards, for example: x.x.x.

Version ARNs have only the first three nodes: <major>.<minor>.<patch>

Build version ARNs have all four nodes, and point to a specific build for a specific version of an object.

imageVersion_owner :: Lens' ImageVersion (Maybe Text)

The owner of the image version.

imageVersion_dateCreated :: Lens' ImageVersion (Maybe Text)

The date on which this specific version of the Image Builder image was created.

imageVersion_name :: Lens' ImageVersion (Maybe Text)

The name of this specific version of an Image Builder image.

imageVersion_version :: Lens' ImageVersion (Maybe Text)

Details for a specific version of an Image Builder image. This version follows the semantic version syntax.

The semantic version has four nodes: <major>.<minor>.<patch>/<build>. You can assign values for the first three, and can filter on all of them.

Assignment: For the first three nodes you can assign any positive integer value, including zero, with an upper limit of 2^30-1, or 1073741823 for each node. Image Builder automatically assigns the build number to the fourth node.

Patterns: You can use any numeric pattern that adheres to the assignment requirements for the nodes that you can assign. For example, you might choose a software version pattern, such as 1.0.0, or a date, such as 2021.01.01.

Filtering: With semantic versioning, you have the flexibility to use wildcards (x) to specify the most recent versions or nodes when selecting the base image or components for your recipe. When you use a wildcard in any node, all nodes to the right of the first wildcard must also be wildcards.

imageVersion_type :: Lens' ImageVersion (Maybe ImageType)

Specifies whether this image is an AMI or a container image.

imageVersion_osVersion :: Lens' ImageVersion (Maybe Text)

The operating system version of the Amazon EC2 build instance. For example, Amazon Linux 2, Ubuntu 18, or Microsoft Windows Server 2019.