Amazonka.ImageBuilder.Types.InfrastructureConfigurationSummary

Description

Documentation

data InfrastructureConfigurationSummary

The infrastructure used when building Amazon EC2 AMIs.

See: newInfrastructureConfigurationSummary smart constructor.

Constructors

InfrastructureConfigurationSummary'

Fields

instanceTypes :: Maybe [Text]

The instance types of the infrastructure configuration.

arn :: Maybe Text

The Amazon Resource Name (ARN) of the infrastructure configuration.

resourceTags :: Maybe (HashMap Text Text)

The tags attached to the image created by Image Builder.

dateUpdated :: Maybe Text

The date on which the infrastructure configuration was last updated.

dateCreated :: Maybe Text

The date on which the infrastructure configuration was created.

name :: Maybe Text

The name of the infrastructure configuration.

instanceProfileName :: Maybe Text

The instance profile of the infrastructure configuration.

description :: Maybe Text

The description of the infrastructure configuration.

tags :: Maybe (HashMap Text Text)

The tags of the infrastructure configuration.

newInfrastructureConfigurationSummary :: InfrastructureConfigurationSummary

Create a value of InfrastructureConfigurationSummary with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:instanceTypes:InfrastructureConfigurationSummary', infrastructureConfigurationSummary_instanceTypes - The instance types of the infrastructure configuration.

$sel:arn:InfrastructureConfigurationSummary', infrastructureConfigurationSummary_arn - The Amazon Resource Name (ARN) of the infrastructure configuration.

$sel:resourceTags:InfrastructureConfigurationSummary', infrastructureConfigurationSummary_resourceTags - The tags attached to the image created by Image Builder.

$sel:dateUpdated:InfrastructureConfigurationSummary', infrastructureConfigurationSummary_dateUpdated - The date on which the infrastructure configuration was last updated.

$sel:dateCreated:InfrastructureConfigurationSummary', infrastructureConfigurationSummary_dateCreated - The date on which the infrastructure configuration was created.

$sel:name:InfrastructureConfigurationSummary', infrastructureConfigurationSummary_name - The name of the infrastructure configuration.

$sel:instanceProfileName:InfrastructureConfigurationSummary', infrastructureConfigurationSummary_instanceProfileName - The instance profile of the infrastructure configuration.

$sel:description:InfrastructureConfigurationSummary', infrastructureConfigurationSummary_description - The description of the infrastructure configuration.

$sel:tags:InfrastructureConfigurationSummary', infrastructureConfigurationSummary_tags - The tags of the infrastructure configuration.

infrastructureConfigurationSummary_instanceTypes :: Lens' InfrastructureConfigurationSummary (Maybe [Text])

The instance types of the infrastructure configuration.

infrastructureConfigurationSummary_arn :: Lens' InfrastructureConfigurationSummary (Maybe Text)

The Amazon Resource Name (ARN) of the infrastructure configuration.

infrastructureConfigurationSummary_resourceTags :: Lens' InfrastructureConfigurationSummary (Maybe (HashMap Text Text))

The tags attached to the image created by Image Builder.

infrastructureConfigurationSummary_dateUpdated :: Lens' InfrastructureConfigurationSummary (Maybe Text)

The date on which the infrastructure configuration was last updated.

infrastructureConfigurationSummary_dateCreated :: Lens' InfrastructureConfigurationSummary (Maybe Text)

The date on which the infrastructure configuration was created.

infrastructureConfigurationSummary_name :: Lens' InfrastructureConfigurationSummary (Maybe Text)

The name of the infrastructure configuration.

infrastructureConfigurationSummary_instanceProfileName :: Lens' InfrastructureConfigurationSummary (Maybe Text)

The instance profile of the infrastructure configuration.

infrastructureConfigurationSummary_description :: Lens' InfrastructureConfigurationSummary (Maybe Text)

The description of the infrastructure configuration.

infrastructureConfigurationSummary_tags :: Lens' InfrastructureConfigurationSummary (Maybe (HashMap Text Text))

The tags of the infrastructure configuration.