Amazonka.ImageBuilder.UpdateImagePipeline

Description

Updates an image pipeline. Image pipelines enable you to automate the creation and distribution of images.

UpdateImagePipeline does not support selective updates for the pipeline. You must specify all of the required properties in the update request, not just the properties that have changed.

Creating a Request

data UpdateImagePipeline

See: newUpdateImagePipeline smart constructor.

Constructors

UpdateImagePipeline'

Fields

status :: Maybe PipelineStatus

The status of the image pipeline.

containerRecipeArn :: Maybe Text

The Amazon Resource Name (ARN) of the container pipeline to update.

imageTestsConfiguration :: Maybe ImageTestsConfiguration

The image test configuration of the image pipeline.

schedule :: Maybe Schedule

The schedule of the image pipeline.

enhancedImageMetadataEnabled :: Maybe Bool

Collects additional information about the image being created, including the operating system (OS) version and package list. This information is used to enhance the overall experience of using EC2 Image Builder. Enabled by default.

distributionConfigurationArn :: Maybe Text

The Amazon Resource Name (ARN) of the distribution configuration that will be used to configure and distribute images updated by this image pipeline.

imageRecipeArn :: Maybe Text

The Amazon Resource Name (ARN) of the image recipe that will be used to configure images updated by this image pipeline.

description :: Maybe Text

The description of the image pipeline.

imagePipelineArn :: Text

The Amazon Resource Name (ARN) of the image pipeline that you want to update.

infrastructureConfigurationArn :: Text

The Amazon Resource Name (ARN) of the infrastructure configuration that will be used to build images updated by this image pipeline.

clientToken :: Text

The idempotency token used to make this request idempotent.

newUpdateImagePipeline

Arguments

:: Text
$sel:imagePipelineArn:UpdateImagePipeline'

-> Text
$sel:infrastructureConfigurationArn:UpdateImagePipeline'

-> Text
$sel:clientToken:UpdateImagePipeline'

-> UpdateImagePipeline

Create a value of UpdateImagePipeline with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:status:UpdateImagePipeline', updateImagePipeline_status - The status of the image pipeline.

$sel:containerRecipeArn:UpdateImagePipeline', updateImagePipeline_containerRecipeArn - The Amazon Resource Name (ARN) of the container pipeline to update.

$sel:imageTestsConfiguration:UpdateImagePipeline', updateImagePipeline_imageTestsConfiguration - The image test configuration of the image pipeline.

$sel:schedule:UpdateImagePipeline', updateImagePipeline_schedule - The schedule of the image pipeline.

$sel:enhancedImageMetadataEnabled:UpdateImagePipeline', updateImagePipeline_enhancedImageMetadataEnabled - Collects additional information about the image being created, including the operating system (OS) version and package list. This information is used to enhance the overall experience of using EC2 Image Builder. Enabled by default.

$sel:distributionConfigurationArn:UpdateImagePipeline', updateImagePipeline_distributionConfigurationArn - The Amazon Resource Name (ARN) of the distribution configuration that will be used to configure and distribute images updated by this image pipeline.

$sel:imageRecipeArn:UpdateImagePipeline', updateImagePipeline_imageRecipeArn - The Amazon Resource Name (ARN) of the image recipe that will be used to configure images updated by this image pipeline.

$sel:description:UpdateImagePipeline', updateImagePipeline_description - The description of the image pipeline.

$sel:imagePipelineArn:UpdateImagePipeline', updateImagePipeline_imagePipelineArn - The Amazon Resource Name (ARN) of the image pipeline that you want to update.

$sel:infrastructureConfigurationArn:UpdateImagePipeline', updateImagePipeline_infrastructureConfigurationArn - The Amazon Resource Name (ARN) of the infrastructure configuration that will be used to build images updated by this image pipeline.

$sel:clientToken:UpdateImagePipeline', updateImagePipeline_clientToken - The idempotency token used to make this request idempotent.

Request Lenses

updateImagePipeline_status :: Lens' UpdateImagePipeline (Maybe PipelineStatus)

The status of the image pipeline.

updateImagePipeline_containerRecipeArn :: Lens' UpdateImagePipeline (Maybe Text)

The Amazon Resource Name (ARN) of the container pipeline to update.

updateImagePipeline_imageTestsConfiguration :: Lens' UpdateImagePipeline (Maybe ImageTestsConfiguration)

The image test configuration of the image pipeline.

updateImagePipeline_schedule :: Lens' UpdateImagePipeline (Maybe Schedule)

The schedule of the image pipeline.

updateImagePipeline_enhancedImageMetadataEnabled :: Lens' UpdateImagePipeline (Maybe Bool)

Collects additional information about the image being created, including the operating system (OS) version and package list. This information is used to enhance the overall experience of using EC2 Image Builder. Enabled by default.

updateImagePipeline_distributionConfigurationArn :: Lens' UpdateImagePipeline (Maybe Text)

The Amazon Resource Name (ARN) of the distribution configuration that will be used to configure and distribute images updated by this image pipeline.

updateImagePipeline_imageRecipeArn :: Lens' UpdateImagePipeline (Maybe Text)

The Amazon Resource Name (ARN) of the image recipe that will be used to configure images updated by this image pipeline.

updateImagePipeline_description :: Lens' UpdateImagePipeline (Maybe Text)

The description of the image pipeline.

updateImagePipeline_imagePipelineArn :: Lens' UpdateImagePipeline Text

The Amazon Resource Name (ARN) of the image pipeline that you want to update.

updateImagePipeline_infrastructureConfigurationArn :: Lens' UpdateImagePipeline Text

The Amazon Resource Name (ARN) of the infrastructure configuration that will be used to build images updated by this image pipeline.

updateImagePipeline_clientToken :: Lens' UpdateImagePipeline Text

The idempotency token used to make this request idempotent.

Destructuring the Response

data UpdateImagePipelineResponse

See: newUpdateImagePipelineResponse smart constructor.

Constructors

UpdateImagePipelineResponse'

Fields

requestId :: Maybe Text

The request ID that uniquely identifies this request.

clientToken :: Maybe Text

The idempotency token used to make this request idempotent.

imagePipelineArn :: Maybe Text

The Amazon Resource Name (ARN) of the image pipeline that was updated by this request.

httpStatus :: Int

The response's http status code.

newUpdateImagePipelineResponse

Arguments

:: Int
$sel:httpStatus:UpdateImagePipelineResponse'

-> UpdateImagePipelineResponse

Create a value of UpdateImagePipelineResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:requestId:UpdateImagePipelineResponse', updateImagePipelineResponse_requestId - The request ID that uniquely identifies this request.

$sel:clientToken:UpdateImagePipeline', updateImagePipelineResponse_clientToken - The idempotency token used to make this request idempotent.

$sel:imagePipelineArn:UpdateImagePipeline', updateImagePipelineResponse_imagePipelineArn - The Amazon Resource Name (ARN) of the image pipeline that was updated by this request.

$sel:httpStatus:UpdateImagePipelineResponse', updateImagePipelineResponse_httpStatus - The response's http status code.

Response Lenses

updateImagePipelineResponse_requestId :: Lens' UpdateImagePipelineResponse (Maybe Text)

The request ID that uniquely identifies this request.

updateImagePipelineResponse_clientToken :: Lens' UpdateImagePipelineResponse (Maybe Text)

The idempotency token used to make this request idempotent.

updateImagePipelineResponse_imagePipelineArn :: Lens' UpdateImagePipelineResponse (Maybe Text)

The Amazon Resource Name (ARN) of the image pipeline that was updated by this request.

updateImagePipelineResponse_httpStatus :: Lens' UpdateImagePipelineResponse Int

The response's http status code.