Amazonka.ImportExport.GetShippingLabel

Description

This operation generates a pre-paid UPS shipping label that you will use to ship your device to AWS for processing.

Creating a Request

data GetShippingLabel

See: newGetShippingLabel smart constructor.

Constructors

GetShippingLabel'

Fields

street3 :: Maybe Text

aPIVersion :: Maybe Text

country :: Maybe Text

stateOrProvince :: Maybe Text

postalCode :: Maybe Text

street2 :: Maybe Text

name :: Maybe Text

company :: Maybe Text

phoneNumber :: Maybe Text

city :: Maybe Text

street1 :: Maybe Text

jobIds :: [Text]

newGetShippingLabel :: GetShippingLabel

Create a value of GetShippingLabel with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:street3:GetShippingLabel', getShippingLabel_street3 - Undocumented member.

$sel:aPIVersion:GetShippingLabel', getShippingLabel_aPIVersion - Undocumented member.

$sel:country:GetShippingLabel', getShippingLabel_country - Undocumented member.

$sel:stateOrProvince:GetShippingLabel', getShippingLabel_stateOrProvince - Undocumented member.

$sel:postalCode:GetShippingLabel', getShippingLabel_postalCode - Undocumented member.

$sel:street2:GetShippingLabel', getShippingLabel_street2 - Undocumented member.

$sel:name:GetShippingLabel', getShippingLabel_name - Undocumented member.

$sel:company:GetShippingLabel', getShippingLabel_company - Undocumented member.

$sel:phoneNumber:GetShippingLabel', getShippingLabel_phoneNumber - Undocumented member.

$sel:city:GetShippingLabel', getShippingLabel_city - Undocumented member.

$sel:street1:GetShippingLabel', getShippingLabel_street1 - Undocumented member.

$sel:jobIds:GetShippingLabel', getShippingLabel_jobIds - Undocumented member.

Request Lenses

getShippingLabel_street3 :: Lens' GetShippingLabel (Maybe Text)

Undocumented member.

getShippingLabel_aPIVersion :: Lens' GetShippingLabel (Maybe Text)

Undocumented member.

getShippingLabel_country :: Lens' GetShippingLabel (Maybe Text)

Undocumented member.

getShippingLabel_stateOrProvince :: Lens' GetShippingLabel (Maybe Text)

Undocumented member.

getShippingLabel_postalCode :: Lens' GetShippingLabel (Maybe Text)

Undocumented member.

getShippingLabel_street2 :: Lens' GetShippingLabel (Maybe Text)

Undocumented member.

getShippingLabel_name :: Lens' GetShippingLabel (Maybe Text)

Undocumented member.

getShippingLabel_company :: Lens' GetShippingLabel (Maybe Text)

Undocumented member.

getShippingLabel_phoneNumber :: Lens' GetShippingLabel (Maybe Text)

Undocumented member.

getShippingLabel_city :: Lens' GetShippingLabel (Maybe Text)

Undocumented member.

getShippingLabel_street1 :: Lens' GetShippingLabel (Maybe Text)

Undocumented member.

getShippingLabel_jobIds :: Lens' GetShippingLabel [Text]

Undocumented member.

Destructuring the Response

data GetShippingLabelResponse

See: newGetShippingLabelResponse smart constructor.

Constructors

GetShippingLabelResponse'

Fields

shippingLabelURL :: Maybe Text

warning :: Maybe Text

httpStatus :: Int

The response's http status code.

newGetShippingLabelResponse

Arguments

:: Int
$sel:httpStatus:GetShippingLabelResponse'

-> GetShippingLabelResponse

Create a value of GetShippingLabelResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:shippingLabelURL:GetShippingLabelResponse', getShippingLabelResponse_shippingLabelURL - Undocumented member.

$sel:warning:GetShippingLabelResponse', getShippingLabelResponse_warning - Undocumented member.

$sel:httpStatus:GetShippingLabelResponse', getShippingLabelResponse_httpStatus - The response's http status code.

Response Lenses

getShippingLabelResponse_shippingLabelURL :: Lens' GetShippingLabelResponse (Maybe Text)

Undocumented member.

getShippingLabelResponse_warning :: Lens' GetShippingLabelResponse (Maybe Text)

Undocumented member.

getShippingLabelResponse_httpStatus :: Lens' GetShippingLabelResponse Int

The response's http status code.