Amazonka.ImportExport.ListJobs

Description

This operation returns the jobs associated with the requester. AWS Import/Export lists the jobs in reverse chronological order based on the date of creation. For example if Job Test1 was created 2009Dec30 and Test2 was created 2010Feb05, the ListJobs operation would return Test2 followed by Test1.

This operation returns paginated results.

Creating a Request

data ListJobs

Input structure for the ListJobs operation.

See: newListJobs smart constructor.

Constructors

ListJobs'

Fields

aPIVersion :: Maybe Text

marker :: Maybe Text

maxJobs :: Maybe Int

newListJobs :: ListJobs

Create a value of ListJobs with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:aPIVersion:ListJobs', listJobs_aPIVersion - Undocumented member.

$sel:marker:ListJobs', listJobs_marker - Undocumented member.

$sel:maxJobs:ListJobs', listJobs_maxJobs - Undocumented member.

Request Lenses

listJobs_aPIVersion :: Lens' ListJobs (Maybe Text)

Undocumented member.

listJobs_marker :: Lens' ListJobs (Maybe Text)

Undocumented member.

listJobs_maxJobs :: Lens' ListJobs (Maybe Int)

Undocumented member.

Destructuring the Response

data ListJobsResponse

Output structure for the ListJobs operation.

See: newListJobsResponse smart constructor.

Constructors

ListJobsResponse'

Fields

jobs :: Maybe [Job]

isTruncated :: Maybe Bool

httpStatus :: Int

The response's http status code.

newListJobsResponse

Arguments

:: Int
$sel:httpStatus:ListJobsResponse'

-> ListJobsResponse

Create a value of ListJobsResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:jobs:ListJobsResponse', listJobsResponse_jobs - Undocumented member.

$sel:isTruncated:ListJobsResponse', listJobsResponse_isTruncated - Undocumented member.

$sel:httpStatus:ListJobsResponse', listJobsResponse_httpStatus - The response's http status code.

Response Lenses

listJobsResponse_jobs :: Lens' ListJobsResponse (Maybe [Job])

Undocumented member.

listJobsResponse_isTruncated :: Lens' ListJobsResponse (Maybe Bool)

Undocumented member.

listJobsResponse_httpStatus :: Lens' ListJobsResponse Int

The response's http status code.