Amazonka.Inspector.PreviewAgents

Description

Previews the agents installed on the EC2 instances that are part of the specified assessment target.

This operation returns paginated results.

Creating a Request

data PreviewAgents

See: newPreviewAgents smart constructor.

Constructors

PreviewAgents'

Fields

nextToken :: Maybe Text

You can use this parameter when paginating results. Set the value of this parameter to null on your first call to the PreviewAgents action. Subsequent calls to the action fill nextToken in the request with the value of NextToken from the previous response to continue listing data.

maxResults :: Maybe Int

You can use this parameter to indicate the maximum number of items you want in the response. The default value is 10. The maximum value is 500.

previewAgentsArn :: Text

The ARN of the assessment target whose agents you want to preview.

newPreviewAgents

Arguments

:: Text
$sel:previewAgentsArn:PreviewAgents'

-> PreviewAgents

Create a value of PreviewAgents with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:nextToken:PreviewAgents', previewAgents_nextToken - You can use this parameter when paginating results. Set the value of this parameter to null on your first call to the PreviewAgents action. Subsequent calls to the action fill nextToken in the request with the value of NextToken from the previous response to continue listing data.

$sel:maxResults:PreviewAgents', previewAgents_maxResults - You can use this parameter to indicate the maximum number of items you want in the response. The default value is 10. The maximum value is 500.

$sel:previewAgentsArn:PreviewAgents', previewAgents_previewAgentsArn - The ARN of the assessment target whose agents you want to preview.

Request Lenses

previewAgents_nextToken :: Lens' PreviewAgents (Maybe Text)

You can use this parameter when paginating results. Set the value of this parameter to null on your first call to the PreviewAgents action. Subsequent calls to the action fill nextToken in the request with the value of NextToken from the previous response to continue listing data.

previewAgents_maxResults :: Lens' PreviewAgents (Maybe Int)

You can use this parameter to indicate the maximum number of items you want in the response. The default value is 10. The maximum value is 500.

previewAgents_previewAgentsArn :: Lens' PreviewAgents Text

The ARN of the assessment target whose agents you want to preview.

Destructuring the Response

data PreviewAgentsResponse

See: newPreviewAgentsResponse smart constructor.

Constructors

PreviewAgentsResponse'

Fields

nextToken :: Maybe Text

When a response is generated, if there is more data to be listed, this parameter is present in the response and contains the value to use for the nextToken parameter in a subsequent pagination request. If there is no more data to be listed, this parameter is set to null.

httpStatus :: Int

The response's http status code.

agentPreviews :: [AgentPreview]

The resulting list of agents.

newPreviewAgentsResponse

Arguments

:: Int
$sel:httpStatus:PreviewAgentsResponse'

-> PreviewAgentsResponse

Create a value of PreviewAgentsResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:nextToken:PreviewAgents', previewAgentsResponse_nextToken - When a response is generated, if there is more data to be listed, this parameter is present in the response and contains the value to use for the nextToken parameter in a subsequent pagination request. If there is no more data to be listed, this parameter is set to null.

$sel:httpStatus:PreviewAgentsResponse', previewAgentsResponse_httpStatus - The response's http status code.

$sel:agentPreviews:PreviewAgentsResponse', previewAgentsResponse_agentPreviews - The resulting list of agents.

Response Lenses

previewAgentsResponse_nextToken :: Lens' PreviewAgentsResponse (Maybe Text)

When a response is generated, if there is more data to be listed, this parameter is present in the response and contains the value to use for the nextToken parameter in a subsequent pagination request. If there is no more data to be listed, this parameter is set to null.

previewAgentsResponse_httpStatus :: Lens' PreviewAgentsResponse Int

The response's http status code.

previewAgentsResponse_agentPreviews :: Lens' PreviewAgentsResponse [AgentPreview]

The resulting list of agents.