Amazonka.Inspector.StartAssessmentRun

Description

Starts the assessment run specified by the ARN of the assessment template. For this API to function properly, you must not exceed the limit of running up to 500 concurrent agents per AWS account.

Creating a Request

data StartAssessmentRun

See: newStartAssessmentRun smart constructor.

Constructors

StartAssessmentRun'

Fields

assessmentRunName :: Maybe Text

You can specify the name for the assessment run. The name must be unique for the assessment template whose ARN is used to start the assessment run.

assessmentTemplateArn :: Text

The ARN of the assessment template of the assessment run that you want to start.

newStartAssessmentRun

Arguments

:: Text
$sel:assessmentTemplateArn:StartAssessmentRun'

-> StartAssessmentRun

Create a value of StartAssessmentRun with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:assessmentRunName:StartAssessmentRun', startAssessmentRun_assessmentRunName - You can specify the name for the assessment run. The name must be unique for the assessment template whose ARN is used to start the assessment run.

$sel:assessmentTemplateArn:StartAssessmentRun', startAssessmentRun_assessmentTemplateArn - The ARN of the assessment template of the assessment run that you want to start.

Request Lenses

startAssessmentRun_assessmentRunName :: Lens' StartAssessmentRun (Maybe Text)

You can specify the name for the assessment run. The name must be unique for the assessment template whose ARN is used to start the assessment run.

startAssessmentRun_assessmentTemplateArn :: Lens' StartAssessmentRun Text

The ARN of the assessment template of the assessment run that you want to start.

Destructuring the Response

data StartAssessmentRunResponse

See: newStartAssessmentRunResponse smart constructor.

Constructors

StartAssessmentRunResponse'

Fields

httpStatus :: Int

The response's http status code.

assessmentRunArn :: Text

The ARN of the assessment run that has been started.

newStartAssessmentRunResponse

Arguments

:: Int
$sel:httpStatus:StartAssessmentRunResponse'

-> Text
$sel:assessmentRunArn:StartAssessmentRunResponse'

-> StartAssessmentRunResponse

Create a value of StartAssessmentRunResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:httpStatus:StartAssessmentRunResponse', startAssessmentRunResponse_httpStatus - The response's http status code.

$sel:assessmentRunArn:StartAssessmentRunResponse', startAssessmentRunResponse_assessmentRunArn - The ARN of the assessment run that has been started.

Response Lenses

startAssessmentRunResponse_httpStatus :: Lens' StartAssessmentRunResponse Int

The response's http status code.

startAssessmentRunResponse_assessmentRunArn :: Lens' StartAssessmentRunResponse Text

The ARN of the assessment run that has been started.