Amazonka.Inspector.Types.AssessmentRun

Description

Documentation

data AssessmentRun

A snapshot of an Amazon Inspector assessment run that contains the findings of the assessment run .

Used as the response element in the DescribeAssessmentRuns action.

See: newAssessmentRun smart constructor.

Constructors

AssessmentRun'

Fields

startedAt :: Maybe POSIX

The time when StartAssessmentRun was called.

completedAt :: Maybe POSIX

The assessment run completion time that corresponds to the rules packages evaluation completion time or failure.

arn :: Text

The ARN of the assessment run.

name :: Text

The auto-generated name for the assessment run.

assessmentTemplateArn :: Text

The ARN of the assessment template that is associated with the assessment run.

state :: AssessmentRunState

The state of the assessment run.

durationInSeconds :: Natural

The duration of the assessment run.

rulesPackageArns :: NonEmpty Text

The rules packages selected for the assessment run.

userAttributesForFindings :: [Attribute]

The user-defined attributes that are assigned to every generated finding.

createdAt :: POSIX

The time when StartAssessmentRun was called.

stateChangedAt :: POSIX

The last time when the assessment run's state changed.

dataCollected :: Bool

A Boolean value (true or false) that specifies whether the process of collecting data from the agents is completed.

stateChanges :: [AssessmentRunStateChange]

A list of the assessment run state changes.

notifications :: [AssessmentRunNotification]

A list of notifications for the event subscriptions. A notification about a particular generated finding is added to this list only once.

findingCounts :: HashMap Severity Int

Provides a total count of generated findings per severity.

newAssessmentRun

Arguments

:: Text
$sel:arn:AssessmentRun'

-> Text
$sel:name:AssessmentRun'

-> Text
$sel:assessmentTemplateArn:AssessmentRun'

-> AssessmentRunState
$sel:state:AssessmentRun'

-> Natural
$sel:durationInSeconds:AssessmentRun'

-> NonEmpty Text
$sel:rulesPackageArns:AssessmentRun'

-> UTCTime
$sel:createdAt:AssessmentRun'

-> UTCTime
$sel:stateChangedAt:AssessmentRun'

-> Bool
$sel:dataCollected:AssessmentRun'

-> AssessmentRun

Create a value of AssessmentRun with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:startedAt:AssessmentRun', assessmentRun_startedAt - The time when StartAssessmentRun was called.

$sel:completedAt:AssessmentRun', assessmentRun_completedAt - The assessment run completion time that corresponds to the rules packages evaluation completion time or failure.

$sel:arn:AssessmentRun', assessmentRun_arn - The ARN of the assessment run.

$sel:name:AssessmentRun', assessmentRun_name - The auto-generated name for the assessment run.

$sel:assessmentTemplateArn:AssessmentRun', assessmentRun_assessmentTemplateArn - The ARN of the assessment template that is associated with the assessment run.

$sel:state:AssessmentRun', assessmentRun_state - The state of the assessment run.

$sel:durationInSeconds:AssessmentRun', assessmentRun_durationInSeconds - The duration of the assessment run.

$sel:rulesPackageArns:AssessmentRun', assessmentRun_rulesPackageArns - The rules packages selected for the assessment run.

$sel:userAttributesForFindings:AssessmentRun', assessmentRun_userAttributesForFindings - The user-defined attributes that are assigned to every generated finding.

$sel:createdAt:AssessmentRun', assessmentRun_createdAt - The time when StartAssessmentRun was called.

$sel:stateChangedAt:AssessmentRun', assessmentRun_stateChangedAt - The last time when the assessment run's state changed.

$sel:dataCollected:AssessmentRun', assessmentRun_dataCollected - A Boolean value (true or false) that specifies whether the process of collecting data from the agents is completed.

$sel:stateChanges:AssessmentRun', assessmentRun_stateChanges - A list of the assessment run state changes.

$sel:notifications:AssessmentRun', assessmentRun_notifications - A list of notifications for the event subscriptions. A notification about a particular generated finding is added to this list only once.

$sel:findingCounts:AssessmentRun', assessmentRun_findingCounts - Provides a total count of generated findings per severity.

assessmentRun_startedAt :: Lens' AssessmentRun (Maybe UTCTime)

The time when StartAssessmentRun was called.

assessmentRun_completedAt :: Lens' AssessmentRun (Maybe UTCTime)

The assessment run completion time that corresponds to the rules packages evaluation completion time or failure.

assessmentRun_arn :: Lens' AssessmentRun Text

The ARN of the assessment run.

assessmentRun_name :: Lens' AssessmentRun Text

The auto-generated name for the assessment run.

assessmentRun_assessmentTemplateArn :: Lens' AssessmentRun Text

The ARN of the assessment template that is associated with the assessment run.

assessmentRun_state :: Lens' AssessmentRun AssessmentRunState

The state of the assessment run.

assessmentRun_durationInSeconds :: Lens' AssessmentRun Natural

The duration of the assessment run.

assessmentRun_rulesPackageArns :: Lens' AssessmentRun (NonEmpty Text)

The rules packages selected for the assessment run.

assessmentRun_userAttributesForFindings :: Lens' AssessmentRun [Attribute]

The user-defined attributes that are assigned to every generated finding.

assessmentRun_createdAt :: Lens' AssessmentRun UTCTime

The time when StartAssessmentRun was called.

assessmentRun_stateChangedAt :: Lens' AssessmentRun UTCTime

The last time when the assessment run's state changed.

assessmentRun_dataCollected :: Lens' AssessmentRun Bool

A Boolean value (true or false) that specifies whether the process of collecting data from the agents is completed.

assessmentRun_stateChanges :: Lens' AssessmentRun [AssessmentRunStateChange]

A list of the assessment run state changes.

assessmentRun_notifications :: Lens' AssessmentRun [AssessmentRunNotification]

A list of notifications for the event subscriptions. A notification about a particular generated finding is added to this list only once.

assessmentRun_findingCounts :: Lens' AssessmentRun (HashMap Severity Int)

Provides a total count of generated findings per severity.