Amazonka.Inspector.Types.AssessmentRunAgent

Description

Documentation

data AssessmentRunAgent

Contains information about an Amazon Inspector agent. This data type is used as a response element in the ListAssessmentRunAgents action.

See: newAssessmentRunAgent smart constructor.

Constructors

AssessmentRunAgent'

Fields

autoScalingGroup :: Maybe Text

The Auto Scaling group of the EC2 instance that is specified by the agent ID.

agentHealthDetails :: Maybe Text

The description for the agent health code.

agentId :: Text

The AWS account of the EC2 instance where the agent is installed.

assessmentRunArn :: Text

The ARN of the assessment run that is associated with the agent.

agentHealth :: AgentHealth

The current health state of the agent.

agentHealthCode :: AgentHealthCode

The detailed health state of the agent.

telemetryMetadata :: [TelemetryMetadata]

The Amazon Inspector application data metrics that are collected by the agent.

newAssessmentRunAgent

Arguments

:: Text
$sel:agentId:AssessmentRunAgent'

-> Text
$sel:assessmentRunArn:AssessmentRunAgent'

-> AgentHealth
$sel:agentHealth:AssessmentRunAgent'

-> AgentHealthCode
$sel:agentHealthCode:AssessmentRunAgent'

-> AssessmentRunAgent

Create a value of AssessmentRunAgent with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:autoScalingGroup:AssessmentRunAgent', assessmentRunAgent_autoScalingGroup - The Auto Scaling group of the EC2 instance that is specified by the agent ID.

$sel:agentHealthDetails:AssessmentRunAgent', assessmentRunAgent_agentHealthDetails - The description for the agent health code.

$sel:agentId:AssessmentRunAgent', assessmentRunAgent_agentId - The AWS account of the EC2 instance where the agent is installed.

$sel:assessmentRunArn:AssessmentRunAgent', assessmentRunAgent_assessmentRunArn - The ARN of the assessment run that is associated with the agent.

$sel:agentHealth:AssessmentRunAgent', assessmentRunAgent_agentHealth - The current health state of the agent.

$sel:agentHealthCode:AssessmentRunAgent', assessmentRunAgent_agentHealthCode - The detailed health state of the agent.

$sel:telemetryMetadata:AssessmentRunAgent', assessmentRunAgent_telemetryMetadata - The Amazon Inspector application data metrics that are collected by the agent.

assessmentRunAgent_autoScalingGroup :: Lens' AssessmentRunAgent (Maybe Text)

The Auto Scaling group of the EC2 instance that is specified by the agent ID.

assessmentRunAgent_agentHealthDetails :: Lens' AssessmentRunAgent (Maybe Text)

The description for the agent health code.

assessmentRunAgent_agentId :: Lens' AssessmentRunAgent Text

The AWS account of the EC2 instance where the agent is installed.

assessmentRunAgent_assessmentRunArn :: Lens' AssessmentRunAgent Text

The ARN of the assessment run that is associated with the agent.

assessmentRunAgent_agentHealth :: Lens' AssessmentRunAgent AgentHealth

The current health state of the agent.

assessmentRunAgent_agentHealthCode :: Lens' AssessmentRunAgent AgentHealthCode

The detailed health state of the agent.

assessmentRunAgent_telemetryMetadata :: Lens' AssessmentRunAgent [TelemetryMetadata]

The Amazon Inspector application data metrics that are collected by the agent.