Amazonka.Inspector.Types.AssessmentTemplate

Description

Documentation

data AssessmentTemplate

Contains information about an Amazon Inspector assessment template. This data type is used as the response element in the DescribeAssessmentTemplates action.

See: newAssessmentTemplate smart constructor.

Constructors

AssessmentTemplate'

Fields

lastAssessmentRunArn :: Maybe Text

The Amazon Resource Name (ARN) of the most recent assessment run associated with this assessment template. This value exists only when the value of assessmentRunCount is greaterpa than zero.

arn :: Text

The ARN of the assessment template.

name :: Text

The name of the assessment template.

assessmentTargetArn :: Text

The ARN of the assessment target that corresponds to this assessment template.

durationInSeconds :: Natural

The duration in seconds specified for this assessment template. The default value is 3600 seconds (one hour). The maximum value is 86400 seconds (one day).

rulesPackageArns :: [Text]

The rules packages that are specified for this assessment template.

userAttributesForFindings :: [Attribute]

The user-defined attributes that are assigned to every generated finding from the assessment run that uses this assessment template.

assessmentRunCount :: Int

The number of existing assessment runs associated with this assessment template. This value can be zero or a positive integer.

createdAt :: POSIX

The time at which the assessment template is created.

newAssessmentTemplate

Arguments

:: Text
$sel:arn:AssessmentTemplate'

-> Text
$sel:name:AssessmentTemplate'

-> Text
$sel:assessmentTargetArn:AssessmentTemplate'

-> Natural
$sel:durationInSeconds:AssessmentTemplate'

-> Int
$sel:assessmentRunCount:AssessmentTemplate'

-> UTCTime
$sel:createdAt:AssessmentTemplate'

-> AssessmentTemplate

Create a value of AssessmentTemplate with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:lastAssessmentRunArn:AssessmentTemplate', assessmentTemplate_lastAssessmentRunArn - The Amazon Resource Name (ARN) of the most recent assessment run associated with this assessment template. This value exists only when the value of assessmentRunCount is greaterpa than zero.

$sel:arn:AssessmentTemplate', assessmentTemplate_arn - The ARN of the assessment template.

$sel:name:AssessmentTemplate', assessmentTemplate_name - The name of the assessment template.

$sel:assessmentTargetArn:AssessmentTemplate', assessmentTemplate_assessmentTargetArn - The ARN of the assessment target that corresponds to this assessment template.

$sel:durationInSeconds:AssessmentTemplate', assessmentTemplate_durationInSeconds - The duration in seconds specified for this assessment template. The default value is 3600 seconds (one hour). The maximum value is 86400 seconds (one day).

$sel:rulesPackageArns:AssessmentTemplate', assessmentTemplate_rulesPackageArns - The rules packages that are specified for this assessment template.

$sel:userAttributesForFindings:AssessmentTemplate', assessmentTemplate_userAttributesForFindings - The user-defined attributes that are assigned to every generated finding from the assessment run that uses this assessment template.

$sel:assessmentRunCount:AssessmentTemplate', assessmentTemplate_assessmentRunCount - The number of existing assessment runs associated with this assessment template. This value can be zero or a positive integer.

$sel:createdAt:AssessmentTemplate', assessmentTemplate_createdAt - The time at which the assessment template is created.

assessmentTemplate_lastAssessmentRunArn :: Lens' AssessmentTemplate (Maybe Text)

The Amazon Resource Name (ARN) of the most recent assessment run associated with this assessment template. This value exists only when the value of assessmentRunCount is greaterpa than zero.

assessmentTemplate_arn :: Lens' AssessmentTemplate Text

The ARN of the assessment template.

assessmentTemplate_name :: Lens' AssessmentTemplate Text

The name of the assessment template.

assessmentTemplate_assessmentTargetArn :: Lens' AssessmentTemplate Text

The ARN of the assessment target that corresponds to this assessment template.

assessmentTemplate_durationInSeconds :: Lens' AssessmentTemplate Natural

The duration in seconds specified for this assessment template. The default value is 3600 seconds (one hour). The maximum value is 86400 seconds (one day).

assessmentTemplate_rulesPackageArns :: Lens' AssessmentTemplate [Text]

The rules packages that are specified for this assessment template.

assessmentTemplate_userAttributesForFindings :: Lens' AssessmentTemplate [Attribute]

The user-defined attributes that are assigned to every generated finding from the assessment run that uses this assessment template.

assessmentTemplate_assessmentRunCount :: Lens' AssessmentTemplate Int

The number of existing assessment runs associated with this assessment template. This value can be zero or a positive integer.

assessmentTemplate_createdAt :: Lens' AssessmentTemplate UTCTime

The time at which the assessment template is created.