Amazonka.Inspector.Types.Finding

Description

Documentation

data Finding

Contains information about an Amazon Inspector finding. This data type is used as the response element in the DescribeFindings action.

See: newFinding smart constructor.

Constructors

Finding'

Fields

service :: Maybe Text

The data element is set to "Inspector".

severity :: Maybe Severity

The finding severity. Values can be set to High, Medium, Low, and Informational.

schemaVersion :: Maybe Natural

The schema version of this data type.

confidence :: Maybe Natural

This data element is currently not used.

assetAttributes :: Maybe AssetAttributes

A collection of attributes of the host from which the finding is generated.

serviceAttributes :: Maybe InspectorServiceAttributes

This data type is used in the Finding data type.

id :: Maybe Text

The ID of the finding.

numericSeverity :: Maybe Double

The numeric value of the finding severity.

assetType :: Maybe AssetType

The type of the host from which the finding is generated.

title :: Maybe Text

The name of the finding.

indicatorOfCompromise :: Maybe Bool

This data element is currently not used.

description :: Maybe Text

The description of the finding.

recommendation :: Maybe Text

The recommendation for the finding.

arn :: Text

The ARN that specifies the finding.

attributes :: [Attribute]

The system-defined attributes for the finding.

userAttributes :: [Attribute]

The user-defined attributes that are assigned to the finding.

createdAt :: POSIX

The time when the finding was generated.

updatedAt :: POSIX

The time when AddAttributesToFindings is called.

newFinding

Arguments

:: Text
$sel:arn:Finding'

-> UTCTime
$sel:createdAt:Finding'

-> UTCTime
$sel:updatedAt:Finding'

-> Finding

Create a value of Finding with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:service:Finding', finding_service - The data element is set to "Inspector".

$sel:severity:Finding', finding_severity - The finding severity. Values can be set to High, Medium, Low, and Informational.

$sel:schemaVersion:Finding', finding_schemaVersion - The schema version of this data type.

$sel:confidence:Finding', finding_confidence - This data element is currently not used.

$sel:assetAttributes:Finding', finding_assetAttributes - A collection of attributes of the host from which the finding is generated.

$sel:serviceAttributes:Finding', finding_serviceAttributes - This data type is used in the Finding data type.

$sel:id:Finding', finding_id - The ID of the finding.

$sel:numericSeverity:Finding', finding_numericSeverity - The numeric value of the finding severity.

$sel:assetType:Finding', finding_assetType - The type of the host from which the finding is generated.

$sel:title:Finding', finding_title - The name of the finding.

$sel:indicatorOfCompromise:Finding', finding_indicatorOfCompromise - This data element is currently not used.

$sel:description:Finding', finding_description - The description of the finding.

$sel:recommendation:Finding', finding_recommendation - The recommendation for the finding.

$sel:arn:Finding', finding_arn - The ARN that specifies the finding.

$sel:attributes:Finding', finding_attributes - The system-defined attributes for the finding.

$sel:userAttributes:Finding', finding_userAttributes - The user-defined attributes that are assigned to the finding.

$sel:createdAt:Finding', finding_createdAt - The time when the finding was generated.

$sel:updatedAt:Finding', finding_updatedAt - The time when AddAttributesToFindings is called.

finding_service :: Lens' Finding (Maybe Text)

The data element is set to "Inspector".

finding_severity :: Lens' Finding (Maybe Severity)

The finding severity. Values can be set to High, Medium, Low, and Informational.

finding_schemaVersion :: Lens' Finding (Maybe Natural)

The schema version of this data type.

finding_confidence :: Lens' Finding (Maybe Natural)

This data element is currently not used.

finding_assetAttributes :: Lens' Finding (Maybe AssetAttributes)

A collection of attributes of the host from which the finding is generated.

finding_serviceAttributes :: Lens' Finding (Maybe InspectorServiceAttributes)

This data type is used in the Finding data type.

finding_id :: Lens' Finding (Maybe Text)

The ID of the finding.

finding_numericSeverity :: Lens' Finding (Maybe Double)

The numeric value of the finding severity.

finding_assetType :: Lens' Finding (Maybe AssetType)

The type of the host from which the finding is generated.

finding_title :: Lens' Finding (Maybe Text)

The name of the finding.

finding_indicatorOfCompromise :: Lens' Finding (Maybe Bool)

This data element is currently not used.

finding_description :: Lens' Finding (Maybe Text)

The description of the finding.

finding_recommendation :: Lens' Finding (Maybe Text)

The recommendation for the finding.

finding_arn :: Lens' Finding Text

The ARN that specifies the finding.

finding_attributes :: Lens' Finding [Attribute]

The system-defined attributes for the finding.

finding_userAttributes :: Lens' Finding [Attribute]

The user-defined attributes that are assigned to the finding.

finding_createdAt :: Lens' Finding UTCTime

The time when the finding was generated.

finding_updatedAt :: Lens' Finding UTCTime

The time when AddAttributesToFindings is called.