Amazonka.IoT.CreateJobTemplate

Description

Creates a job template.

Requires permission to access the CreateJobTemplate action.

Creating a Request

data CreateJobTemplate

See: newCreateJobTemplate smart constructor.

Constructors

CreateJobTemplate'

Fields

jobExecutionsRolloutConfig :: Maybe JobExecutionsRolloutConfig

jobArn :: Maybe Text

The ARN of the job to use as the basis for the job template.

documentSource :: Maybe Text

An S3 link to the job document to use in the template. Required if you don't specify a value for document.

If the job document resides in an S3 bucket, you must use a placeholder link when specifying the document.

The placeholder link is of the following form:

${aws:iot:s3-presigned-url:https://s3.amazonaws.com/bucket/key}

where bucket is your bucket name and key is the object in the bucket to which you are linking.

abortConfig :: Maybe AbortConfig

presignedUrlConfig :: Maybe PresignedUrlConfig

document :: Maybe Text

The job document. Required if you don't specify a value for documentSource.

timeoutConfig :: Maybe TimeoutConfig

tags :: Maybe [Tag]

Metadata that can be used to manage the job template.

jobTemplateId :: Text

A unique identifier for the job template. We recommend using a UUID. Alpha-numeric characters, "-", and "_" are valid for use here.

description :: Text

A description of the job document.

newCreateJobTemplate

Arguments

:: Text
$sel:jobTemplateId:CreateJobTemplate'

-> Text
$sel:description:CreateJobTemplate'

-> CreateJobTemplate

Create a value of CreateJobTemplate with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:jobExecutionsRolloutConfig:CreateJobTemplate', createJobTemplate_jobExecutionsRolloutConfig - Undocumented member.

$sel:jobArn:CreateJobTemplate', createJobTemplate_jobArn - The ARN of the job to use as the basis for the job template.

$sel:documentSource:CreateJobTemplate', createJobTemplate_documentSource - An S3 link to the job document to use in the template. Required if you don't specify a value for document.

If the job document resides in an S3 bucket, you must use a placeholder link when specifying the document.

The placeholder link is of the following form:

${aws:iot:s3-presigned-url:https://s3.amazonaws.com/bucket/key}

where bucket is your bucket name and key is the object in the bucket to which you are linking.

$sel:abortConfig:CreateJobTemplate', createJobTemplate_abortConfig - Undocumented member.

$sel:presignedUrlConfig:CreateJobTemplate', createJobTemplate_presignedUrlConfig - Undocumented member.

$sel:document:CreateJobTemplate', createJobTemplate_document - The job document. Required if you don't specify a value for documentSource.

$sel:timeoutConfig:CreateJobTemplate', createJobTemplate_timeoutConfig - Undocumented member.

$sel:tags:CreateJobTemplate', createJobTemplate_tags - Metadata that can be used to manage the job template.

$sel:jobTemplateId:CreateJobTemplate', createJobTemplate_jobTemplateId - A unique identifier for the job template. We recommend using a UUID. Alpha-numeric characters, "-", and "_" are valid for use here.

$sel:description:CreateJobTemplate', createJobTemplate_description - A description of the job document.

Request Lenses

createJobTemplate_jobExecutionsRolloutConfig :: Lens' CreateJobTemplate (Maybe JobExecutionsRolloutConfig)

Undocumented member.

createJobTemplate_jobArn :: Lens' CreateJobTemplate (Maybe Text)

The ARN of the job to use as the basis for the job template.

createJobTemplate_documentSource :: Lens' CreateJobTemplate (Maybe Text)

An S3 link to the job document to use in the template. Required if you don't specify a value for document.

If the job document resides in an S3 bucket, you must use a placeholder link when specifying the document.

The placeholder link is of the following form:

${aws:iot:s3-presigned-url:https://s3.amazonaws.com/bucket/key}

where bucket is your bucket name and key is the object in the bucket to which you are linking.

createJobTemplate_abortConfig :: Lens' CreateJobTemplate (Maybe AbortConfig)

Undocumented member.

createJobTemplate_presignedUrlConfig :: Lens' CreateJobTemplate (Maybe PresignedUrlConfig)

Undocumented member.

createJobTemplate_document :: Lens' CreateJobTemplate (Maybe Text)

The job document. Required if you don't specify a value for documentSource.

createJobTemplate_timeoutConfig :: Lens' CreateJobTemplate (Maybe TimeoutConfig)

Undocumented member.

createJobTemplate_tags :: Lens' CreateJobTemplate (Maybe [Tag])

Metadata that can be used to manage the job template.

createJobTemplate_jobTemplateId :: Lens' CreateJobTemplate Text

A unique identifier for the job template. We recommend using a UUID. Alpha-numeric characters, "-", and "_" are valid for use here.

createJobTemplate_description :: Lens' CreateJobTemplate Text

A description of the job document.

Destructuring the Response

data CreateJobTemplateResponse

See: newCreateJobTemplateResponse smart constructor.

Constructors

CreateJobTemplateResponse'

Fields

jobTemplateId :: Maybe Text

The unique identifier of the job template.

jobTemplateArn :: Maybe Text

The ARN of the job template.

httpStatus :: Int

The response's http status code.

newCreateJobTemplateResponse

Arguments

:: Int
$sel:httpStatus:CreateJobTemplateResponse'

-> CreateJobTemplateResponse

Create a value of CreateJobTemplateResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:jobTemplateId:CreateJobTemplate', createJobTemplateResponse_jobTemplateId - The unique identifier of the job template.

$sel:jobTemplateArn:CreateJobTemplateResponse', createJobTemplateResponse_jobTemplateArn - The ARN of the job template.

$sel:httpStatus:CreateJobTemplateResponse', createJobTemplateResponse_httpStatus - The response's http status code.

Response Lenses

createJobTemplateResponse_jobTemplateId :: Lens' CreateJobTemplateResponse (Maybe Text)

The unique identifier of the job template.

createJobTemplateResponse_jobTemplateArn :: Lens' CreateJobTemplateResponse (Maybe Text)

The ARN of the job template.

createJobTemplateResponse_httpStatus :: Lens' CreateJobTemplateResponse Int

The response's http status code.