Amazonka.IoT.CreateKeysAndCertificate

Description

Creates a 2048-bit RSA key pair and issues an X.509 certificate using the issued public key. You can also call CreateKeysAndCertificate over MQTT from a device, for more information, see Provisioning MQTT API.

Note This is the only time IoT issues the private key for this certificate, so it is important to keep it in a secure location.

Requires permission to access the CreateKeysAndCertificate action.

Creating a Request

data CreateKeysAndCertificate

The input for the CreateKeysAndCertificate operation.

Requires permission to access the CreateKeysAndCertificateRequest action.

See: newCreateKeysAndCertificate smart constructor.

Constructors

CreateKeysAndCertificate'

Fields

setAsActive :: Maybe Bool

Specifies whether the certificate is active.

newCreateKeysAndCertificate :: CreateKeysAndCertificate

Create a value of CreateKeysAndCertificate with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:setAsActive:CreateKeysAndCertificate', createKeysAndCertificate_setAsActive - Specifies whether the certificate is active.

Request Lenses

createKeysAndCertificate_setAsActive :: Lens' CreateKeysAndCertificate (Maybe Bool)

Specifies whether the certificate is active.

Destructuring the Response

data CreateKeysAndCertificateResponse

The output of the CreateKeysAndCertificate operation.

See: newCreateKeysAndCertificateResponse smart constructor.

Constructors

CreateKeysAndCertificateResponse'

Fields

keyPair :: Maybe KeyPair

The generated key pair.

certificatePem :: Maybe Text

The certificate data, in PEM format.

certificateArn :: Maybe Text

The ARN of the certificate.

certificateId :: Maybe Text

The ID of the certificate. IoT issues a default subject name for the certificate (for example, IoT Certificate).

httpStatus :: Int

The response's http status code.

newCreateKeysAndCertificateResponse

Arguments

:: Int
$sel:httpStatus:CreateKeysAndCertificateResponse'

-> CreateKeysAndCertificateResponse

Create a value of CreateKeysAndCertificateResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:keyPair:CreateKeysAndCertificateResponse', createKeysAndCertificateResponse_keyPair - The generated key pair.

$sel:certificatePem:CreateKeysAndCertificateResponse', createKeysAndCertificateResponse_certificatePem - The certificate data, in PEM format.

$sel:certificateArn:CreateKeysAndCertificateResponse', createKeysAndCertificateResponse_certificateArn - The ARN of the certificate.

$sel:certificateId:CreateKeysAndCertificateResponse', createKeysAndCertificateResponse_certificateId - The ID of the certificate. IoT issues a default subject name for the certificate (for example, IoT Certificate).

$sel:httpStatus:CreateKeysAndCertificateResponse', createKeysAndCertificateResponse_httpStatus - The response's http status code.

Response Lenses

createKeysAndCertificateResponse_keyPair :: Lens' CreateKeysAndCertificateResponse (Maybe KeyPair)

The generated key pair.

createKeysAndCertificateResponse_certificatePem :: Lens' CreateKeysAndCertificateResponse (Maybe Text)

The certificate data, in PEM format.

createKeysAndCertificateResponse_certificateArn :: Lens' CreateKeysAndCertificateResponse (Maybe Text)

The ARN of the certificate.

createKeysAndCertificateResponse_certificateId :: Lens' CreateKeysAndCertificateResponse (Maybe Text)

The ID of the certificate. IoT issues a default subject name for the certificate (for example, IoT Certificate).

createKeysAndCertificateResponse_httpStatus :: Lens' CreateKeysAndCertificateResponse Int

The response's http status code.