Amazonka.IoT.DeleteCACertificate

Description

Deletes a registered CA certificate.

Requires permission to access the DeleteCACertificate action.

Creating a Request

data DeleteCACertificate

Input for the DeleteCACertificate operation.

See: newDeleteCACertificate smart constructor.

Constructors

DeleteCACertificate'

Fields

certificateId :: Text

The ID of the certificate to delete. (The last part of the certificate ARN contains the certificate ID.)

newDeleteCACertificate

Arguments

:: Text
$sel:certificateId:DeleteCACertificate'

-> DeleteCACertificate

Create a value of DeleteCACertificate with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:certificateId:DeleteCACertificate', deleteCACertificate_certificateId - The ID of the certificate to delete. (The last part of the certificate ARN contains the certificate ID.)

Request Lenses

deleteCACertificate_certificateId :: Lens' DeleteCACertificate Text

The ID of the certificate to delete. (The last part of the certificate ARN contains the certificate ID.)

Destructuring the Response

data DeleteCACertificateResponse

The output for the DeleteCACertificate operation.

See: newDeleteCACertificateResponse smart constructor.

Constructors

DeleteCACertificateResponse'

Fields

httpStatus :: Int

The response's http status code.

newDeleteCACertificateResponse

Arguments

:: Int
$sel:httpStatus:DeleteCACertificateResponse'

-> DeleteCACertificateResponse

Create a value of DeleteCACertificateResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:httpStatus:DeleteCACertificateResponse', deleteCACertificateResponse_httpStatus - The response's http status code.

Response Lenses

deleteCACertificateResponse_httpStatus :: Lens' DeleteCACertificateResponse Int

The response's http status code.