Amazonka.IoT.DeleteCertificate

Description

Deletes the specified certificate.

A certificate cannot be deleted if it has a policy or IoT thing attached to it or if its status is set to ACTIVE. To delete a certificate, first use the DetachPolicy action to detach all policies. Next, use the UpdateCertificate action to set the certificate to the INACTIVE status.

Requires permission to access the DeleteCertificate action.

Creating a Request

data DeleteCertificate

The input for the DeleteCertificate operation.

See: newDeleteCertificate smart constructor.

Constructors

DeleteCertificate'

Fields

forceDelete :: Maybe Bool

Forces the deletion of a certificate if it is inactive and is not attached to an IoT thing.

certificateId :: Text

The ID of the certificate. (The last part of the certificate ARN contains the certificate ID.)

newDeleteCertificate

Arguments

:: Text
$sel:certificateId:DeleteCertificate'

-> DeleteCertificate

Create a value of DeleteCertificate with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:forceDelete:DeleteCertificate', deleteCertificate_forceDelete - Forces the deletion of a certificate if it is inactive and is not attached to an IoT thing.

$sel:certificateId:DeleteCertificate', deleteCertificate_certificateId - The ID of the certificate. (The last part of the certificate ARN contains the certificate ID.)

Request Lenses

deleteCertificate_forceDelete :: Lens' DeleteCertificate (Maybe Bool)

Forces the deletion of a certificate if it is inactive and is not attached to an IoT thing.

deleteCertificate_certificateId :: Lens' DeleteCertificate Text

The ID of the certificate. (The last part of the certificate ARN contains the certificate ID.)

Destructuring the Response

data DeleteCertificateResponse

See: newDeleteCertificateResponse smart constructor.

Constructors

DeleteCertificateResponse'

newDeleteCertificateResponse :: DeleteCertificateResponse

Create a value of DeleteCertificateResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.