Amazonka.IoT.DeletePolicy

Description

Deletes the specified policy.

A policy cannot be deleted if it has non-default versions or it is attached to any certificate.

To delete a policy, use the DeletePolicyVersion action to delete all non-default versions of the policy; use the DetachPolicy action to detach the policy from any certificate; and then use the DeletePolicy action to delete the policy.

When a policy is deleted using DeletePolicy, its default version is deleted with it.

Because of the distributed nature of Amazon Web Services, it can take up to five minutes after a policy is detached before it's ready to be deleted.

Requires permission to access the DeletePolicy action.

Creating a Request

data DeletePolicy

The input for the DeletePolicy operation.

See: newDeletePolicy smart constructor.

Constructors

DeletePolicy'

Fields

policyName :: Text

The name of the policy to delete.

newDeletePolicy

Arguments

:: Text
$sel:policyName:DeletePolicy'

-> DeletePolicy

Create a value of DeletePolicy with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:policyName:DeletePolicy', deletePolicy_policyName - The name of the policy to delete.

Request Lenses

deletePolicy_policyName :: Lens' DeletePolicy Text

The name of the policy to delete.

Destructuring the Response

data DeletePolicyResponse

See: newDeletePolicyResponse smart constructor.

Constructors

DeletePolicyResponse'

newDeletePolicyResponse :: DeletePolicyResponse

Create a value of DeletePolicyResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.