Amazonka.IoT.DescribeAuditTask

Description

Gets information about a Device Defender audit.

Requires permission to access the DescribeAuditTask action.

Creating a Request

data DescribeAuditTask

See: newDescribeAuditTask smart constructor.

Constructors

DescribeAuditTask'

Fields

taskId :: Text

The ID of the audit whose information you want to get.

newDescribeAuditTask

Arguments

:: Text
$sel:taskId:DescribeAuditTask'

-> DescribeAuditTask

Create a value of DescribeAuditTask with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:taskId:DescribeAuditTask', describeAuditTask_taskId - The ID of the audit whose information you want to get.

Request Lenses

describeAuditTask_taskId :: Lens' DescribeAuditTask Text

The ID of the audit whose information you want to get.

Destructuring the Response

data DescribeAuditTaskResponse

See: newDescribeAuditTaskResponse smart constructor.

Constructors

DescribeAuditTaskResponse'

Fields

auditDetails :: Maybe (HashMap Text AuditCheckDetails)

Detailed information about each check performed during this audit.

taskType :: Maybe AuditTaskType

The type of audit: "ON_DEMAND_AUDIT_TASK" or "SCHEDULED_AUDIT_TASK".

taskStartTime :: Maybe POSIX

The time the audit started.

taskStatistics :: Maybe TaskStatistics

Statistical information about the audit.

scheduledAuditName :: Maybe Text

The name of the scheduled audit (only if the audit was a scheduled audit).

taskStatus :: Maybe AuditTaskStatus

The status of the audit: one of "IN_PROGRESS", "COMPLETED", "FAILED", or "CANCELED".

httpStatus :: Int

The response's http status code.

newDescribeAuditTaskResponse

Arguments

:: Int
$sel:httpStatus:DescribeAuditTaskResponse'

-> DescribeAuditTaskResponse

Create a value of DescribeAuditTaskResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:auditDetails:DescribeAuditTaskResponse', describeAuditTaskResponse_auditDetails - Detailed information about each check performed during this audit.

$sel:taskType:DescribeAuditTaskResponse', describeAuditTaskResponse_taskType - The type of audit: "ON_DEMAND_AUDIT_TASK" or "SCHEDULED_AUDIT_TASK".

$sel:taskStartTime:DescribeAuditTaskResponse', describeAuditTaskResponse_taskStartTime - The time the audit started.

$sel:taskStatistics:DescribeAuditTaskResponse', describeAuditTaskResponse_taskStatistics - Statistical information about the audit.

$sel:scheduledAuditName:DescribeAuditTaskResponse', describeAuditTaskResponse_scheduledAuditName - The name of the scheduled audit (only if the audit was a scheduled audit).

$sel:taskStatus:DescribeAuditTaskResponse', describeAuditTaskResponse_taskStatus - The status of the audit: one of "IN_PROGRESS", "COMPLETED", "FAILED", or "CANCELED".

$sel:httpStatus:DescribeAuditTaskResponse', describeAuditTaskResponse_httpStatus - The response's http status code.

Response Lenses

describeAuditTaskResponse_auditDetails :: Lens' DescribeAuditTaskResponse (Maybe (HashMap Text AuditCheckDetails))

Detailed information about each check performed during this audit.

describeAuditTaskResponse_taskType :: Lens' DescribeAuditTaskResponse (Maybe AuditTaskType)

The type of audit: "ON_DEMAND_AUDIT_TASK" or "SCHEDULED_AUDIT_TASK".

describeAuditTaskResponse_taskStartTime :: Lens' DescribeAuditTaskResponse (Maybe UTCTime)

The time the audit started.

describeAuditTaskResponse_taskStatistics :: Lens' DescribeAuditTaskResponse (Maybe TaskStatistics)

Statistical information about the audit.

describeAuditTaskResponse_scheduledAuditName :: Lens' DescribeAuditTaskResponse (Maybe Text)

The name of the scheduled audit (only if the audit was a scheduled audit).

describeAuditTaskResponse_taskStatus :: Lens' DescribeAuditTaskResponse (Maybe AuditTaskStatus)

The status of the audit: one of "IN_PROGRESS", "COMPLETED", "FAILED", or "CANCELED".

describeAuditTaskResponse_httpStatus :: Lens' DescribeAuditTaskResponse Int

The response's http status code.