Amazonka.IoT.DescribeCACertificate

Description

Describes a registered CA certificate.

Requires permission to access the DescribeCACertificate action.

Creating a Request

data DescribeCACertificate

The input for the DescribeCACertificate operation.

See: newDescribeCACertificate smart constructor.

Constructors

DescribeCACertificate'

Fields

certificateId :: Text

The CA certificate identifier.

newDescribeCACertificate

Arguments

:: Text
$sel:certificateId:DescribeCACertificate'

-> DescribeCACertificate

Create a value of DescribeCACertificate with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:certificateId:DescribeCACertificate', describeCACertificate_certificateId - The CA certificate identifier.

Request Lenses

describeCACertificate_certificateId :: Lens' DescribeCACertificate Text

The CA certificate identifier.

Destructuring the Response

data DescribeCACertificateResponse

The output from the DescribeCACertificate operation.

See: newDescribeCACertificateResponse smart constructor.

Constructors

DescribeCACertificateResponse'

Fields

certificateDescription :: Maybe CACertificateDescription

The CA certificate description.

registrationConfig :: Maybe RegistrationConfig

Information about the registration configuration.

httpStatus :: Int

The response's http status code.

newDescribeCACertificateResponse

Arguments

:: Int
$sel:httpStatus:DescribeCACertificateResponse'

-> DescribeCACertificateResponse

Create a value of DescribeCACertificateResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:certificateDescription:DescribeCACertificateResponse', describeCACertificateResponse_certificateDescription - The CA certificate description.

$sel:registrationConfig:DescribeCACertificateResponse', describeCACertificateResponse_registrationConfig - Information about the registration configuration.

$sel:httpStatus:DescribeCACertificateResponse', describeCACertificateResponse_httpStatus - The response's http status code.

Response Lenses

describeCACertificateResponse_certificateDescription :: Lens' DescribeCACertificateResponse (Maybe CACertificateDescription)

The CA certificate description.

describeCACertificateResponse_registrationConfig :: Lens' DescribeCACertificateResponse (Maybe RegistrationConfig)

Information about the registration configuration.

describeCACertificateResponse_httpStatus :: Lens' DescribeCACertificateResponse Int

The response's http status code.