Amazonka.IoT.DescribeSecurityProfile

Description

Gets information about a Device Defender security profile.

Requires permission to access the DescribeSecurityProfile action.

Creating a Request

data DescribeSecurityProfile

See: newDescribeSecurityProfile smart constructor.

Constructors

DescribeSecurityProfile'

Fields

securityProfileName :: Text

The name of the security profile whose information you want to get.

newDescribeSecurityProfile

Arguments

:: Text
$sel:securityProfileName:DescribeSecurityProfile'

-> DescribeSecurityProfile

Create a value of DescribeSecurityProfile with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:securityProfileName:DescribeSecurityProfile', describeSecurityProfile_securityProfileName - The name of the security profile whose information you want to get.

Request Lenses

describeSecurityProfile_securityProfileName :: Lens' DescribeSecurityProfile Text

The name of the security profile whose information you want to get.

Destructuring the Response

data DescribeSecurityProfileResponse

See: newDescribeSecurityProfileResponse smart constructor.

Constructors

DescribeSecurityProfileResponse'

Fields

alertTargets :: Maybe (HashMap AlertTargetType AlertTarget)

Where the alerts are sent. (Alerts are always sent to the console.)

additionalMetricsToRetainV2 :: Maybe [MetricToRetain]

A list of metrics whose data is retained (stored). By default, data is retained for any metric used in the profile's behaviors, but it is also retained for any metric specified here.

behaviors :: Maybe [Behavior]

Specifies the behaviors that, when violated by a device (thing), cause an alert.

lastModifiedDate :: Maybe POSIX

The time the security profile was last modified.

version :: Maybe Integer

The version of the security profile. A new version is generated whenever the security profile is updated.

securityProfileName :: Maybe Text

The name of the security profile.

creationDate :: Maybe POSIX

The time the security profile was created.

additionalMetricsToRetain :: Maybe [Text]

/Please use DescribeSecurityProfileResponse$additionalMetricsToRetainV2 instead./

A list of metrics whose data is retained (stored). By default, data is retained for any metric used in the profile's behaviors, but it is also retained for any metric specified here.

securityProfileArn :: Maybe Text

The ARN of the security profile.

securityProfileDescription :: Maybe Text

A description of the security profile (associated with the security profile when it was created or updated).

httpStatus :: Int

The response's http status code.

newDescribeSecurityProfileResponse

Arguments

:: Int
$sel:httpStatus:DescribeSecurityProfileResponse'

-> DescribeSecurityProfileResponse

Create a value of DescribeSecurityProfileResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:alertTargets:DescribeSecurityProfileResponse', describeSecurityProfileResponse_alertTargets - Where the alerts are sent. (Alerts are always sent to the console.)

$sel:additionalMetricsToRetainV2:DescribeSecurityProfileResponse', describeSecurityProfileResponse_additionalMetricsToRetainV2 - A list of metrics whose data is retained (stored). By default, data is retained for any metric used in the profile's behaviors, but it is also retained for any metric specified here.

$sel:behaviors:DescribeSecurityProfileResponse', describeSecurityProfileResponse_behaviors - Specifies the behaviors that, when violated by a device (thing), cause an alert.

$sel:lastModifiedDate:DescribeSecurityProfileResponse', describeSecurityProfileResponse_lastModifiedDate - The time the security profile was last modified.

$sel:version:DescribeSecurityProfileResponse', describeSecurityProfileResponse_version - The version of the security profile. A new version is generated whenever the security profile is updated.

$sel:securityProfileName:DescribeSecurityProfile', describeSecurityProfileResponse_securityProfileName - The name of the security profile.

$sel:creationDate:DescribeSecurityProfileResponse', describeSecurityProfileResponse_creationDate - The time the security profile was created.

$sel:additionalMetricsToRetain:DescribeSecurityProfileResponse', describeSecurityProfileResponse_additionalMetricsToRetain - /Please use DescribeSecurityProfileResponse$additionalMetricsToRetainV2 instead./

A list of metrics whose data is retained (stored). By default, data is retained for any metric used in the profile's behaviors, but it is also retained for any metric specified here.

$sel:securityProfileArn:DescribeSecurityProfileResponse', describeSecurityProfileResponse_securityProfileArn - The ARN of the security profile.

$sel:securityProfileDescription:DescribeSecurityProfileResponse', describeSecurityProfileResponse_securityProfileDescription - A description of the security profile (associated with the security profile when it was created or updated).

$sel:httpStatus:DescribeSecurityProfileResponse', describeSecurityProfileResponse_httpStatus - The response's http status code.

Response Lenses

describeSecurityProfileResponse_alertTargets :: Lens' DescribeSecurityProfileResponse (Maybe (HashMap AlertTargetType AlertTarget))

Where the alerts are sent. (Alerts are always sent to the console.)

describeSecurityProfileResponse_additionalMetricsToRetainV2 :: Lens' DescribeSecurityProfileResponse (Maybe [MetricToRetain])

A list of metrics whose data is retained (stored). By default, data is retained for any metric used in the profile's behaviors, but it is also retained for any metric specified here.

describeSecurityProfileResponse_behaviors :: Lens' DescribeSecurityProfileResponse (Maybe [Behavior])

Specifies the behaviors that, when violated by a device (thing), cause an alert.

describeSecurityProfileResponse_lastModifiedDate :: Lens' DescribeSecurityProfileResponse (Maybe UTCTime)

The time the security profile was last modified.

describeSecurityProfileResponse_version :: Lens' DescribeSecurityProfileResponse (Maybe Integer)

The version of the security profile. A new version is generated whenever the security profile is updated.

describeSecurityProfileResponse_securityProfileName :: Lens' DescribeSecurityProfileResponse (Maybe Text)

The name of the security profile.

describeSecurityProfileResponse_creationDate :: Lens' DescribeSecurityProfileResponse (Maybe UTCTime)

The time the security profile was created.

describeSecurityProfileResponse_additionalMetricsToRetain :: Lens' DescribeSecurityProfileResponse (Maybe [Text])

/Please use DescribeSecurityProfileResponse$additionalMetricsToRetainV2 instead./

A list of metrics whose data is retained (stored). By default, data is retained for any metric used in the profile's behaviors, but it is also retained for any metric specified here.

describeSecurityProfileResponse_securityProfileArn :: Lens' DescribeSecurityProfileResponse (Maybe Text)

The ARN of the security profile.

describeSecurityProfileResponse_securityProfileDescription :: Lens' DescribeSecurityProfileResponse (Maybe Text)

A description of the security profile (associated with the security profile when it was created or updated).

describeSecurityProfileResponse_httpStatus :: Lens' DescribeSecurityProfileResponse Int

The response's http status code.