Amazonka.IoT.DescribeThingRegistrationTask

Description

Describes a bulk thing provisioning task.

Requires permission to access the DescribeThingRegistrationTask action.

Creating a Request

data DescribeThingRegistrationTask

See: newDescribeThingRegistrationTask smart constructor.

Constructors

DescribeThingRegistrationTask'

Fields

taskId :: Text

The task ID.

newDescribeThingRegistrationTask

Arguments

:: Text
$sel:taskId:DescribeThingRegistrationTask'

-> DescribeThingRegistrationTask

Create a value of DescribeThingRegistrationTask with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:taskId:DescribeThingRegistrationTask', describeThingRegistrationTask_taskId - The task ID.

Request Lenses

describeThingRegistrationTask_taskId :: Lens' DescribeThingRegistrationTask Text

The task ID.

Destructuring the Response

data DescribeThingRegistrationTaskResponse

See: newDescribeThingRegistrationTaskResponse smart constructor.

Constructors

DescribeThingRegistrationTaskResponse'

Fields

status :: Maybe TaskStatus

The status of the bulk thing provisioning task.

lastModifiedDate :: Maybe POSIX

The date when the task was last modified.

inputFileKey :: Maybe Text

The input file key.

taskId :: Maybe Text

The task ID.

creationDate :: Maybe POSIX

The task creation date.

percentageProgress :: Maybe Natural

The progress of the bulk provisioning task expressed as a percentage.

templateBody :: Maybe Text

The task's template.

successCount :: Maybe Int

The number of things successfully provisioned.

message :: Maybe Text

The message.

failureCount :: Maybe Int

The number of things that failed to be provisioned.

inputFileBucket :: Maybe Text

The S3 bucket that contains the input file.

roleArn :: Maybe Text

The role ARN that grants access to the input file bucket.

httpStatus :: Int

The response's http status code.

newDescribeThingRegistrationTaskResponse

Arguments

:: Int
$sel:httpStatus:DescribeThingRegistrationTaskResponse'

-> DescribeThingRegistrationTaskResponse

Create a value of DescribeThingRegistrationTaskResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:status:DescribeThingRegistrationTaskResponse', describeThingRegistrationTaskResponse_status - The status of the bulk thing provisioning task.

$sel:lastModifiedDate:DescribeThingRegistrationTaskResponse', describeThingRegistrationTaskResponse_lastModifiedDate - The date when the task was last modified.

$sel:inputFileKey:DescribeThingRegistrationTaskResponse', describeThingRegistrationTaskResponse_inputFileKey - The input file key.

$sel:taskId:DescribeThingRegistrationTask', describeThingRegistrationTaskResponse_taskId - The task ID.

$sel:creationDate:DescribeThingRegistrationTaskResponse', describeThingRegistrationTaskResponse_creationDate - The task creation date.

$sel:percentageProgress:DescribeThingRegistrationTaskResponse', describeThingRegistrationTaskResponse_percentageProgress - The progress of the bulk provisioning task expressed as a percentage.

$sel:templateBody:DescribeThingRegistrationTaskResponse', describeThingRegistrationTaskResponse_templateBody - The task's template.

$sel:successCount:DescribeThingRegistrationTaskResponse', describeThingRegistrationTaskResponse_successCount - The number of things successfully provisioned.

$sel:message:DescribeThingRegistrationTaskResponse', describeThingRegistrationTaskResponse_message - The message.

$sel:failureCount:DescribeThingRegistrationTaskResponse', describeThingRegistrationTaskResponse_failureCount - The number of things that failed to be provisioned.

$sel:inputFileBucket:DescribeThingRegistrationTaskResponse', describeThingRegistrationTaskResponse_inputFileBucket - The S3 bucket that contains the input file.

$sel:roleArn:DescribeThingRegistrationTaskResponse', describeThingRegistrationTaskResponse_roleArn - The role ARN that grants access to the input file bucket.

$sel:httpStatus:DescribeThingRegistrationTaskResponse', describeThingRegistrationTaskResponse_httpStatus - The response's http status code.

Response Lenses

describeThingRegistrationTaskResponse_status :: Lens' DescribeThingRegistrationTaskResponse (Maybe TaskStatus)

The status of the bulk thing provisioning task.

describeThingRegistrationTaskResponse_lastModifiedDate :: Lens' DescribeThingRegistrationTaskResponse (Maybe UTCTime)

The date when the task was last modified.

describeThingRegistrationTaskResponse_inputFileKey :: Lens' DescribeThingRegistrationTaskResponse (Maybe Text)

The input file key.

describeThingRegistrationTaskResponse_taskId :: Lens' DescribeThingRegistrationTaskResponse (Maybe Text)

The task ID.

describeThingRegistrationTaskResponse_creationDate :: Lens' DescribeThingRegistrationTaskResponse (Maybe UTCTime)

The task creation date.

describeThingRegistrationTaskResponse_percentageProgress :: Lens' DescribeThingRegistrationTaskResponse (Maybe Natural)

The progress of the bulk provisioning task expressed as a percentage.

describeThingRegistrationTaskResponse_templateBody :: Lens' DescribeThingRegistrationTaskResponse (Maybe Text)

The task's template.

describeThingRegistrationTaskResponse_successCount :: Lens' DescribeThingRegistrationTaskResponse (Maybe Int)

The number of things successfully provisioned.

describeThingRegistrationTaskResponse_message :: Lens' DescribeThingRegistrationTaskResponse (Maybe Text)

The message.

describeThingRegistrationTaskResponse_failureCount :: Lens' DescribeThingRegistrationTaskResponse (Maybe Int)

The number of things that failed to be provisioned.

describeThingRegistrationTaskResponse_inputFileBucket :: Lens' DescribeThingRegistrationTaskResponse (Maybe Text)

The S3 bucket that contains the input file.

describeThingRegistrationTaskResponse_roleArn :: Lens' DescribeThingRegistrationTaskResponse (Maybe Text)

The role ARN that grants access to the input file bucket.

describeThingRegistrationTaskResponse_httpStatus :: Lens' DescribeThingRegistrationTaskResponse Int

The response's http status code.