Amazonka.IoT.GetPolicy

Description

Gets information about the specified policy with the policy document of the default version.

Requires permission to access the GetPolicy action.

Creating a Request

data GetPolicy

The input for the GetPolicy operation.

See: newGetPolicy smart constructor.

Constructors

GetPolicy'

Fields

policyName :: Text

The name of the policy.

newGetPolicy

Arguments

:: Text
$sel:policyName:GetPolicy'

-> GetPolicy

Create a value of GetPolicy with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:policyName:GetPolicy', getPolicy_policyName - The name of the policy.

Request Lenses

getPolicy_policyName :: Lens' GetPolicy Text

The name of the policy.

Destructuring the Response

data GetPolicyResponse

The output from the GetPolicy operation.

See: newGetPolicyResponse smart constructor.

Constructors

GetPolicyResponse'

Fields

lastModifiedDate :: Maybe POSIX

The date the policy was last modified.

policyName :: Maybe Text

The policy name.

policyDocument :: Maybe Text

The JSON document that describes the policy.

defaultVersionId :: Maybe Text

The default policy version ID.

policyArn :: Maybe Text

The policy ARN.

creationDate :: Maybe POSIX

The date the policy was created.

generationId :: Maybe Text

The generation ID of the policy.

httpStatus :: Int

The response's http status code.

newGetPolicyResponse

Arguments

:: Int
$sel:httpStatus:GetPolicyResponse'

-> GetPolicyResponse

Create a value of GetPolicyResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:lastModifiedDate:GetPolicyResponse', getPolicyResponse_lastModifiedDate - The date the policy was last modified.

$sel:policyName:GetPolicy', getPolicyResponse_policyName - The policy name.

$sel:policyDocument:GetPolicyResponse', getPolicyResponse_policyDocument - The JSON document that describes the policy.

$sel:defaultVersionId:GetPolicyResponse', getPolicyResponse_defaultVersionId - The default policy version ID.

$sel:policyArn:GetPolicyResponse', getPolicyResponse_policyArn - The policy ARN.

$sel:creationDate:GetPolicyResponse', getPolicyResponse_creationDate - The date the policy was created.

$sel:generationId:GetPolicyResponse', getPolicyResponse_generationId - The generation ID of the policy.

$sel:httpStatus:GetPolicyResponse', getPolicyResponse_httpStatus - The response's http status code.

Response Lenses

getPolicyResponse_lastModifiedDate :: Lens' GetPolicyResponse (Maybe UTCTime)

The date the policy was last modified.

getPolicyResponse_policyName :: Lens' GetPolicyResponse (Maybe Text)

The policy name.

getPolicyResponse_policyDocument :: Lens' GetPolicyResponse (Maybe Text)

The JSON document that describes the policy.

getPolicyResponse_defaultVersionId :: Lens' GetPolicyResponse (Maybe Text)

The default policy version ID.

getPolicyResponse_policyArn :: Lens' GetPolicyResponse (Maybe Text)

The policy ARN.

getPolicyResponse_creationDate :: Lens' GetPolicyResponse (Maybe UTCTime)

The date the policy was created.

getPolicyResponse_generationId :: Lens' GetPolicyResponse (Maybe Text)

The generation ID of the policy.

getPolicyResponse_httpStatus :: Lens' GetPolicyResponse Int

The response's http status code.