Amazonka.IoT.GetPolicyVersion

Description

Gets information about the specified policy version.

Requires permission to access the GetPolicyVersion action.

Creating a Request

data GetPolicyVersion

The input for the GetPolicyVersion operation.

See: newGetPolicyVersion smart constructor.

Constructors

GetPolicyVersion'

Fields

policyName :: Text

The name of the policy.

policyVersionId :: Text

The policy version ID.

newGetPolicyVersion

Arguments

:: Text
$sel:policyName:GetPolicyVersion'

-> Text
$sel:policyVersionId:GetPolicyVersion'

-> GetPolicyVersion

Create a value of GetPolicyVersion with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:policyName:GetPolicyVersion', getPolicyVersion_policyName - The name of the policy.

$sel:policyVersionId:GetPolicyVersion', getPolicyVersion_policyVersionId - The policy version ID.

Request Lenses

getPolicyVersion_policyName :: Lens' GetPolicyVersion Text

The name of the policy.

getPolicyVersion_policyVersionId :: Lens' GetPolicyVersion Text

The policy version ID.

Destructuring the Response

data GetPolicyVersionResponse

The output from the GetPolicyVersion operation.

See: newGetPolicyVersionResponse smart constructor.

Constructors

GetPolicyVersionResponse'

Fields

lastModifiedDate :: Maybe POSIX

The date the policy was last modified.

policyName :: Maybe Text

The policy name.

policyDocument :: Maybe Text

The JSON document that describes the policy.

policyVersionId :: Maybe Text

The policy version ID.

policyArn :: Maybe Text

The policy ARN.

creationDate :: Maybe POSIX

The date the policy was created.

generationId :: Maybe Text

The generation ID of the policy version.

isDefaultVersion :: Maybe Bool

Specifies whether the policy version is the default.

httpStatus :: Int

The response's http status code.

newGetPolicyVersionResponse

Arguments

:: Int
$sel:httpStatus:GetPolicyVersionResponse'

-> GetPolicyVersionResponse

Create a value of GetPolicyVersionResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:lastModifiedDate:GetPolicyVersionResponse', getPolicyVersionResponse_lastModifiedDate - The date the policy was last modified.

$sel:policyName:GetPolicyVersion', getPolicyVersionResponse_policyName - The policy name.

$sel:policyDocument:GetPolicyVersionResponse', getPolicyVersionResponse_policyDocument - The JSON document that describes the policy.

$sel:policyVersionId:GetPolicyVersion', getPolicyVersionResponse_policyVersionId - The policy version ID.

$sel:policyArn:GetPolicyVersionResponse', getPolicyVersionResponse_policyArn - The policy ARN.

$sel:creationDate:GetPolicyVersionResponse', getPolicyVersionResponse_creationDate - The date the policy was created.

$sel:generationId:GetPolicyVersionResponse', getPolicyVersionResponse_generationId - The generation ID of the policy version.

$sel:isDefaultVersion:GetPolicyVersionResponse', getPolicyVersionResponse_isDefaultVersion - Specifies whether the policy version is the default.

$sel:httpStatus:GetPolicyVersionResponse', getPolicyVersionResponse_httpStatus - The response's http status code.

Response Lenses

getPolicyVersionResponse_lastModifiedDate :: Lens' GetPolicyVersionResponse (Maybe UTCTime)

The date the policy was last modified.

getPolicyVersionResponse_policyName :: Lens' GetPolicyVersionResponse (Maybe Text)

The policy name.

getPolicyVersionResponse_policyDocument :: Lens' GetPolicyVersionResponse (Maybe Text)

The JSON document that describes the policy.

getPolicyVersionResponse_policyVersionId :: Lens' GetPolicyVersionResponse (Maybe Text)

The policy version ID.

getPolicyVersionResponse_policyArn :: Lens' GetPolicyVersionResponse (Maybe Text)

The policy ARN.

getPolicyVersionResponse_creationDate :: Lens' GetPolicyVersionResponse (Maybe UTCTime)

The date the policy was created.

getPolicyVersionResponse_generationId :: Lens' GetPolicyVersionResponse (Maybe Text)

The generation ID of the policy version.

getPolicyVersionResponse_isDefaultVersion :: Lens' GetPolicyVersionResponse (Maybe Bool)

Specifies whether the policy version is the default.

getPolicyVersionResponse_httpStatus :: Lens' GetPolicyVersionResponse Int

The response's http status code.