Amazonka.IoT.GetStatistics

Description

Returns the count, average, sum, minimum, maximum, sum of squares, variance, and standard deviation for the specified aggregated field. If the aggregation field is of type String, only the count statistic is returned.

Requires permission to access the GetStatistics action.

Creating a Request

data GetStatistics

See: newGetStatistics smart constructor.

Constructors

GetStatistics'

Fields

queryVersion :: Maybe Text

The version of the query used to search.

aggregationField :: Maybe Text

The aggregation field name.

indexName :: Maybe Text

The name of the index to search. The default value is AWS_Things.

queryString :: Text

The query used to search. You can specify "*" for the query string to get the count of all indexed things in your Amazon Web Services account.

newGetStatistics

Arguments

:: Text
$sel:queryString:GetStatistics'

-> GetStatistics

Create a value of GetStatistics with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:queryVersion:GetStatistics', getStatistics_queryVersion - The version of the query used to search.

$sel:aggregationField:GetStatistics', getStatistics_aggregationField - The aggregation field name.

$sel:indexName:GetStatistics', getStatistics_indexName - The name of the index to search. The default value is AWS_Things.

$sel:queryString:GetStatistics', getStatistics_queryString - The query used to search. You can specify "*" for the query string to get the count of all indexed things in your Amazon Web Services account.

Request Lenses

getStatistics_queryVersion :: Lens' GetStatistics (Maybe Text)

The version of the query used to search.

getStatistics_aggregationField :: Lens' GetStatistics (Maybe Text)

The aggregation field name.

getStatistics_indexName :: Lens' GetStatistics (Maybe Text)

The name of the index to search. The default value is AWS_Things.

getStatistics_queryString :: Lens' GetStatistics Text

The query used to search. You can specify "*" for the query string to get the count of all indexed things in your Amazon Web Services account.

Destructuring the Response

data GetStatisticsResponse

See: newGetStatisticsResponse smart constructor.

Constructors

GetStatisticsResponse'

Fields

statistics :: Maybe Statistics

The statistics returned by the Fleet Indexing service based on the query and aggregation field.

httpStatus :: Int

The response's http status code.

newGetStatisticsResponse

Arguments

:: Int
$sel:httpStatus:GetStatisticsResponse'

-> GetStatisticsResponse

Create a value of GetStatisticsResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:statistics:GetStatisticsResponse', getStatisticsResponse_statistics - The statistics returned by the Fleet Indexing service based on the query and aggregation field.

$sel:httpStatus:GetStatisticsResponse', getStatisticsResponse_httpStatus - The response's http status code.

Response Lenses

getStatisticsResponse_statistics :: Lens' GetStatisticsResponse (Maybe Statistics)

The statistics returned by the Fleet Indexing service based on the query and aggregation field.

getStatisticsResponse_httpStatus :: Lens' GetStatisticsResponse Int

The response's http status code.