Amazonka.IoT.ListAuditFindings

Description

Lists the findings (results) of a Device Defender audit or of the audits performed during a specified time period. (Findings are retained for 90 days.)

Requires permission to access the ListAuditFindings action.

This operation returns paginated results.

Creating a Request

data ListAuditFindings

See: newListAuditFindings smart constructor.

Constructors

ListAuditFindings'

Fields

startTime :: Maybe POSIX

A filter to limit results to those found after the specified time. You must specify either the startTime and endTime or the taskId, but not both.

taskId :: Maybe Text

A filter to limit results to the audit with the specified ID. You must specify either the taskId or the startTime and endTime, but not both.

checkName :: Maybe Text

A filter to limit results to the findings for the specified audit check.

listSuppressedFindings :: Maybe Bool

Boolean flag indicating whether only the suppressed findings or the unsuppressed findings should be listed. If this parameter isn't provided, the response will list both suppressed and unsuppressed findings.

nextToken :: Maybe Text

The token for the next set of results.

endTime :: Maybe POSIX

A filter to limit results to those found before the specified time. You must specify either the startTime and endTime or the taskId, but not both.

maxResults :: Maybe Natural

The maximum number of results to return at one time. The default is 25.

resourceIdentifier :: Maybe ResourceIdentifier

Information identifying the noncompliant resource.

newListAuditFindings :: ListAuditFindings

Create a value of ListAuditFindings with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:startTime:ListAuditFindings', listAuditFindings_startTime - A filter to limit results to those found after the specified time. You must specify either the startTime and endTime or the taskId, but not both.

$sel:taskId:ListAuditFindings', listAuditFindings_taskId - A filter to limit results to the audit with the specified ID. You must specify either the taskId or the startTime and endTime, but not both.

$sel:checkName:ListAuditFindings', listAuditFindings_checkName - A filter to limit results to the findings for the specified audit check.

$sel:listSuppressedFindings:ListAuditFindings', listAuditFindings_listSuppressedFindings - Boolean flag indicating whether only the suppressed findings or the unsuppressed findings should be listed. If this parameter isn't provided, the response will list both suppressed and unsuppressed findings.

$sel:nextToken:ListAuditFindings', listAuditFindings_nextToken - The token for the next set of results.

$sel:endTime:ListAuditFindings', listAuditFindings_endTime - A filter to limit results to those found before the specified time. You must specify either the startTime and endTime or the taskId, but not both.

$sel:maxResults:ListAuditFindings', listAuditFindings_maxResults - The maximum number of results to return at one time. The default is 25.

$sel:resourceIdentifier:ListAuditFindings', listAuditFindings_resourceIdentifier - Information identifying the noncompliant resource.

Request Lenses

listAuditFindings_startTime :: Lens' ListAuditFindings (Maybe UTCTime)

A filter to limit results to those found after the specified time. You must specify either the startTime and endTime or the taskId, but not both.

listAuditFindings_taskId :: Lens' ListAuditFindings (Maybe Text)

A filter to limit results to the audit with the specified ID. You must specify either the taskId or the startTime and endTime, but not both.

listAuditFindings_checkName :: Lens' ListAuditFindings (Maybe Text)

A filter to limit results to the findings for the specified audit check.

listAuditFindings_listSuppressedFindings :: Lens' ListAuditFindings (Maybe Bool)

Boolean flag indicating whether only the suppressed findings or the unsuppressed findings should be listed. If this parameter isn't provided, the response will list both suppressed and unsuppressed findings.

listAuditFindings_nextToken :: Lens' ListAuditFindings (Maybe Text)

The token for the next set of results.

listAuditFindings_endTime :: Lens' ListAuditFindings (Maybe UTCTime)

A filter to limit results to those found before the specified time. You must specify either the startTime and endTime or the taskId, but not both.

listAuditFindings_maxResults :: Lens' ListAuditFindings (Maybe Natural)

The maximum number of results to return at one time. The default is 25.

listAuditFindings_resourceIdentifier :: Lens' ListAuditFindings (Maybe ResourceIdentifier)

Information identifying the noncompliant resource.

Destructuring the Response

data ListAuditFindingsResponse

See: newListAuditFindingsResponse smart constructor.

Constructors

ListAuditFindingsResponse'

Fields

nextToken :: Maybe Text

A token that can be used to retrieve the next set of results, or null if there are no additional results.

findings :: Maybe [AuditFinding]

The findings (results) of the audit.

httpStatus :: Int

The response's http status code.

newListAuditFindingsResponse

Arguments

:: Int
$sel:httpStatus:ListAuditFindingsResponse'

-> ListAuditFindingsResponse

Create a value of ListAuditFindingsResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:nextToken:ListAuditFindings', listAuditFindingsResponse_nextToken - A token that can be used to retrieve the next set of results, or null if there are no additional results.

$sel:findings:ListAuditFindingsResponse', listAuditFindingsResponse_findings - The findings (results) of the audit.

$sel:httpStatus:ListAuditFindingsResponse', listAuditFindingsResponse_httpStatus - The response's http status code.

Response Lenses

listAuditFindingsResponse_nextToken :: Lens' ListAuditFindingsResponse (Maybe Text)

A token that can be used to retrieve the next set of results, or null if there are no additional results.

listAuditFindingsResponse_findings :: Lens' ListAuditFindingsResponse (Maybe [AuditFinding])

The findings (results) of the audit.

listAuditFindingsResponse_httpStatus :: Lens' ListAuditFindingsResponse Int

The response's http status code.