Amazonka.IoT.ListAuditMitigationActionsTasks

Description

Gets a list of audit mitigation action tasks that match the specified filters.

Requires permission to access the ListAuditMitigationActionsTasks action.

This operation returns paginated results.

Creating a Request

data ListAuditMitigationActionsTasks

See: newListAuditMitigationActionsTasks smart constructor.

Constructors

ListAuditMitigationActionsTasks'

Fields

auditTaskId :: Maybe Text

Specify this filter to limit results to tasks that were applied to results for a specific audit.

nextToken :: Maybe Text

The token for the next set of results.

findingId :: Maybe Text

Specify this filter to limit results to tasks that were applied to a specific audit finding.

maxResults :: Maybe Natural

The maximum number of results to return at one time. The default is 25.

taskStatus :: Maybe AuditMitigationActionsTaskStatus

Specify this filter to limit results to tasks that are in a specific state.

startTime :: POSIX

Specify this filter to limit results to tasks that began on or after a specific date and time.

endTime :: POSIX

Specify this filter to limit results to tasks that were completed or canceled on or before a specific date and time.

newListAuditMitigationActionsTasks

Arguments

:: UTCTime
$sel:startTime:ListAuditMitigationActionsTasks'

-> UTCTime
$sel:endTime:ListAuditMitigationActionsTasks'

-> ListAuditMitigationActionsTasks

Create a value of ListAuditMitigationActionsTasks with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:auditTaskId:ListAuditMitigationActionsTasks', listAuditMitigationActionsTasks_auditTaskId - Specify this filter to limit results to tasks that were applied to results for a specific audit.

$sel:nextToken:ListAuditMitigationActionsTasks', listAuditMitigationActionsTasks_nextToken - The token for the next set of results.

$sel:findingId:ListAuditMitigationActionsTasks', listAuditMitigationActionsTasks_findingId - Specify this filter to limit results to tasks that were applied to a specific audit finding.

$sel:maxResults:ListAuditMitigationActionsTasks', listAuditMitigationActionsTasks_maxResults - The maximum number of results to return at one time. The default is 25.

$sel:taskStatus:ListAuditMitigationActionsTasks', listAuditMitigationActionsTasks_taskStatus - Specify this filter to limit results to tasks that are in a specific state.

$sel:startTime:ListAuditMitigationActionsTasks', listAuditMitigationActionsTasks_startTime - Specify this filter to limit results to tasks that began on or after a specific date and time.

$sel:endTime:ListAuditMitigationActionsTasks', listAuditMitigationActionsTasks_endTime - Specify this filter to limit results to tasks that were completed or canceled on or before a specific date and time.

Request Lenses

listAuditMitigationActionsTasks_auditTaskId :: Lens' ListAuditMitigationActionsTasks (Maybe Text)

Specify this filter to limit results to tasks that were applied to results for a specific audit.

listAuditMitigationActionsTasks_nextToken :: Lens' ListAuditMitigationActionsTasks (Maybe Text)

The token for the next set of results.

listAuditMitigationActionsTasks_findingId :: Lens' ListAuditMitigationActionsTasks (Maybe Text)

Specify this filter to limit results to tasks that were applied to a specific audit finding.

listAuditMitigationActionsTasks_maxResults :: Lens' ListAuditMitigationActionsTasks (Maybe Natural)

The maximum number of results to return at one time. The default is 25.

listAuditMitigationActionsTasks_taskStatus :: Lens' ListAuditMitigationActionsTasks (Maybe AuditMitigationActionsTaskStatus)

Specify this filter to limit results to tasks that are in a specific state.

listAuditMitigationActionsTasks_startTime :: Lens' ListAuditMitigationActionsTasks UTCTime

Specify this filter to limit results to tasks that began on or after a specific date and time.

listAuditMitigationActionsTasks_endTime :: Lens' ListAuditMitigationActionsTasks UTCTime

Specify this filter to limit results to tasks that were completed or canceled on or before a specific date and time.

Destructuring the Response

data ListAuditMitigationActionsTasksResponse

See: newListAuditMitigationActionsTasksResponse smart constructor.

Constructors

ListAuditMitigationActionsTasksResponse'

Fields

tasks :: Maybe [AuditMitigationActionsTaskMetadata]

The collection of audit mitigation tasks that matched the filter criteria.

nextToken :: Maybe Text

The token for the next set of results.

httpStatus :: Int

The response's http status code.

newListAuditMitigationActionsTasksResponse

Arguments

:: Int
$sel:httpStatus:ListAuditMitigationActionsTasksResponse'

-> ListAuditMitigationActionsTasksResponse

Create a value of ListAuditMitigationActionsTasksResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:tasks:ListAuditMitigationActionsTasksResponse', listAuditMitigationActionsTasksResponse_tasks - The collection of audit mitigation tasks that matched the filter criteria.

$sel:nextToken:ListAuditMitigationActionsTasks', listAuditMitigationActionsTasksResponse_nextToken - The token for the next set of results.

$sel:httpStatus:ListAuditMitigationActionsTasksResponse', listAuditMitigationActionsTasksResponse_httpStatus - The response's http status code.

Response Lenses

listAuditMitigationActionsTasksResponse_tasks :: Lens' ListAuditMitigationActionsTasksResponse (Maybe [AuditMitigationActionsTaskMetadata])

The collection of audit mitigation tasks that matched the filter criteria.

listAuditMitigationActionsTasksResponse_nextToken :: Lens' ListAuditMitigationActionsTasksResponse (Maybe Text)

The token for the next set of results.

listAuditMitigationActionsTasksResponse_httpStatus :: Lens' ListAuditMitigationActionsTasksResponse Int

The response's http status code.