Amazonka.IoT.ListRoleAliases

Description

Lists the role aliases registered in your account.

Requires permission to access the ListRoleAliases action.

This operation returns paginated results.

Creating a Request

data ListRoleAliases

See: newListRoleAliases smart constructor.

Constructors

ListRoleAliases'

Fields

marker :: Maybe Text

A marker used to get the next set of results.

ascendingOrder :: Maybe Bool

Return the list of role aliases in ascending alphabetical order.

pageSize :: Maybe Natural

The maximum number of results to return at one time.

newListRoleAliases :: ListRoleAliases

Create a value of ListRoleAliases with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:marker:ListRoleAliases', listRoleAliases_marker - A marker used to get the next set of results.

$sel:ascendingOrder:ListRoleAliases', listRoleAliases_ascendingOrder - Return the list of role aliases in ascending alphabetical order.

$sel:pageSize:ListRoleAliases', listRoleAliases_pageSize - The maximum number of results to return at one time.

Request Lenses

listRoleAliases_marker :: Lens' ListRoleAliases (Maybe Text)

A marker used to get the next set of results.

listRoleAliases_ascendingOrder :: Lens' ListRoleAliases (Maybe Bool)

Return the list of role aliases in ascending alphabetical order.

listRoleAliases_pageSize :: Lens' ListRoleAliases (Maybe Natural)

The maximum number of results to return at one time.

Destructuring the Response

data ListRoleAliasesResponse

See: newListRoleAliasesResponse smart constructor.

Constructors

ListRoleAliasesResponse'

Fields

roleAliases :: Maybe [Text]

The role aliases.

nextMarker :: Maybe Text

A marker used to get the next set of results.

httpStatus :: Int

The response's http status code.

newListRoleAliasesResponse

Arguments

:: Int
$sel:httpStatus:ListRoleAliasesResponse'

-> ListRoleAliasesResponse

Create a value of ListRoleAliasesResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:roleAliases:ListRoleAliasesResponse', listRoleAliasesResponse_roleAliases - The role aliases.

$sel:nextMarker:ListRoleAliasesResponse', listRoleAliasesResponse_nextMarker - A marker used to get the next set of results.

$sel:httpStatus:ListRoleAliasesResponse', listRoleAliasesResponse_httpStatus - The response's http status code.

Response Lenses

listRoleAliasesResponse_roleAliases :: Lens' ListRoleAliasesResponse (Maybe [Text])

The role aliases.

listRoleAliasesResponse_nextMarker :: Lens' ListRoleAliasesResponse (Maybe Text)

A marker used to get the next set of results.

listRoleAliasesResponse_httpStatus :: Lens' ListRoleAliasesResponse Int

The response's http status code.