Amazonka.IoT.ListSecurityProfiles

Description

Lists the Device Defender security profiles you've created. You can filter security profiles by dimension or custom metric.

Requires permission to access the ListSecurityProfiles action.

dimensionName and metricName cannot be used in the same request.

This operation returns paginated results.

Creating a Request

data ListSecurityProfiles

See: newListSecurityProfiles smart constructor.

Constructors

ListSecurityProfiles'

Fields

metricName :: Maybe Text

The name of the custom metric. Cannot be used with dimensionName.

nextToken :: Maybe Text

The token for the next set of results.

dimensionName :: Maybe Text

A filter to limit results to the security profiles that use the defined dimension. Cannot be used with metricName

maxResults :: Maybe Natural

The maximum number of results to return at one time.

newListSecurityProfiles :: ListSecurityProfiles

Create a value of ListSecurityProfiles with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:metricName:ListSecurityProfiles', listSecurityProfiles_metricName - The name of the custom metric. Cannot be used with dimensionName.

$sel:nextToken:ListSecurityProfiles', listSecurityProfiles_nextToken - The token for the next set of results.

$sel:dimensionName:ListSecurityProfiles', listSecurityProfiles_dimensionName - A filter to limit results to the security profiles that use the defined dimension. Cannot be used with metricName

$sel:maxResults:ListSecurityProfiles', listSecurityProfiles_maxResults - The maximum number of results to return at one time.

Request Lenses

listSecurityProfiles_metricName :: Lens' ListSecurityProfiles (Maybe Text)

The name of the custom metric. Cannot be used with dimensionName.

listSecurityProfiles_nextToken :: Lens' ListSecurityProfiles (Maybe Text)

The token for the next set of results.

listSecurityProfiles_dimensionName :: Lens' ListSecurityProfiles (Maybe Text)

A filter to limit results to the security profiles that use the defined dimension. Cannot be used with metricName

listSecurityProfiles_maxResults :: Lens' ListSecurityProfiles (Maybe Natural)

The maximum number of results to return at one time.

Destructuring the Response

data ListSecurityProfilesResponse

See: newListSecurityProfilesResponse smart constructor.

Constructors

ListSecurityProfilesResponse'

Fields

nextToken :: Maybe Text

A token that can be used to retrieve the next set of results, or null if there are no additional results.

securityProfileIdentifiers :: Maybe [SecurityProfileIdentifier]

A list of security profile identifiers (names and ARNs).

httpStatus :: Int

The response's http status code.

newListSecurityProfilesResponse

Arguments

:: Int
$sel:httpStatus:ListSecurityProfilesResponse'

-> ListSecurityProfilesResponse

Create a value of ListSecurityProfilesResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:nextToken:ListSecurityProfiles', listSecurityProfilesResponse_nextToken - A token that can be used to retrieve the next set of results, or null if there are no additional results.

$sel:securityProfileIdentifiers:ListSecurityProfilesResponse', listSecurityProfilesResponse_securityProfileIdentifiers - A list of security profile identifiers (names and ARNs).

$sel:httpStatus:ListSecurityProfilesResponse', listSecurityProfilesResponse_httpStatus - The response's http status code.

Response Lenses

listSecurityProfilesResponse_nextToken :: Lens' ListSecurityProfilesResponse (Maybe Text)

A token that can be used to retrieve the next set of results, or null if there are no additional results.

listSecurityProfilesResponse_securityProfileIdentifiers :: Lens' ListSecurityProfilesResponse (Maybe [SecurityProfileIdentifier])

A list of security profile identifiers (names and ARNs).

listSecurityProfilesResponse_httpStatus :: Lens' ListSecurityProfilesResponse Int

The response's http status code.