Amazonka.IoT.ListViolationEvents

Description

Lists the Device Defender security profile violations discovered during the given time period. You can use filters to limit the results to those alerts issued for a particular security profile, behavior, or thing (device).

Requires permission to access the ListViolationEvents action.

This operation returns paginated results.

Creating a Request

data ListViolationEvents

See: newListViolationEvents smart constructor.

Constructors

ListViolationEvents'

Fields

nextToken :: Maybe Text

The token for the next set of results.

listSuppressedAlerts :: Maybe Bool

A list of all suppressed alerts.

behaviorCriteriaType :: Maybe BehaviorCriteriaType

The criteria for a behavior.

securityProfileName :: Maybe Text

A filter to limit results to those alerts generated by the specified security profile.

thingName :: Maybe Text

A filter to limit results to those alerts caused by the specified thing.

verificationState :: Maybe VerificationState

The verification state of the violation (detect alarm).

maxResults :: Maybe Natural

The maximum number of results to return at one time.

startTime :: POSIX

The start time for the alerts to be listed.

endTime :: POSIX

The end time for the alerts to be listed.

newListViolationEvents

Arguments

:: UTCTime
$sel:startTime:ListViolationEvents'

-> UTCTime
$sel:endTime:ListViolationEvents'

-> ListViolationEvents

Create a value of ListViolationEvents with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:nextToken:ListViolationEvents', listViolationEvents_nextToken - The token for the next set of results.

$sel:listSuppressedAlerts:ListViolationEvents', listViolationEvents_listSuppressedAlerts - A list of all suppressed alerts.

$sel:behaviorCriteriaType:ListViolationEvents', listViolationEvents_behaviorCriteriaType - The criteria for a behavior.

$sel:securityProfileName:ListViolationEvents', listViolationEvents_securityProfileName - A filter to limit results to those alerts generated by the specified security profile.

$sel:thingName:ListViolationEvents', listViolationEvents_thingName - A filter to limit results to those alerts caused by the specified thing.

$sel:verificationState:ListViolationEvents', listViolationEvents_verificationState - The verification state of the violation (detect alarm).

$sel:maxResults:ListViolationEvents', listViolationEvents_maxResults - The maximum number of results to return at one time.

$sel:startTime:ListViolationEvents', listViolationEvents_startTime - The start time for the alerts to be listed.

$sel:endTime:ListViolationEvents', listViolationEvents_endTime - The end time for the alerts to be listed.

Request Lenses

listViolationEvents_nextToken :: Lens' ListViolationEvents (Maybe Text)

The token for the next set of results.

listViolationEvents_listSuppressedAlerts :: Lens' ListViolationEvents (Maybe Bool)

A list of all suppressed alerts.

listViolationEvents_behaviorCriteriaType :: Lens' ListViolationEvents (Maybe BehaviorCriteriaType)

The criteria for a behavior.

listViolationEvents_securityProfileName :: Lens' ListViolationEvents (Maybe Text)

A filter to limit results to those alerts generated by the specified security profile.

listViolationEvents_thingName :: Lens' ListViolationEvents (Maybe Text)

A filter to limit results to those alerts caused by the specified thing.

listViolationEvents_verificationState :: Lens' ListViolationEvents (Maybe VerificationState)

The verification state of the violation (detect alarm).

listViolationEvents_maxResults :: Lens' ListViolationEvents (Maybe Natural)

The maximum number of results to return at one time.

listViolationEvents_startTime :: Lens' ListViolationEvents UTCTime

The start time for the alerts to be listed.

listViolationEvents_endTime :: Lens' ListViolationEvents UTCTime

The end time for the alerts to be listed.

Destructuring the Response

data ListViolationEventsResponse

See: newListViolationEventsResponse smart constructor.

Constructors

ListViolationEventsResponse'

Fields

violationEvents :: Maybe [ViolationEvent]

The security profile violation alerts issued for this account during the given time period, potentially filtered by security profile, behavior violated, or thing (device) violating.

nextToken :: Maybe Text

A token that can be used to retrieve the next set of results, or null if there are no additional results.

httpStatus :: Int

The response's http status code.

newListViolationEventsResponse

Arguments

:: Int
$sel:httpStatus:ListViolationEventsResponse'

-> ListViolationEventsResponse

Create a value of ListViolationEventsResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:violationEvents:ListViolationEventsResponse', listViolationEventsResponse_violationEvents - The security profile violation alerts issued for this account during the given time period, potentially filtered by security profile, behavior violated, or thing (device) violating.

$sel:nextToken:ListViolationEvents', listViolationEventsResponse_nextToken - A token that can be used to retrieve the next set of results, or null if there are no additional results.

$sel:httpStatus:ListViolationEventsResponse', listViolationEventsResponse_httpStatus - The response's http status code.

Response Lenses

listViolationEventsResponse_violationEvents :: Lens' ListViolationEventsResponse (Maybe [ViolationEvent])

The security profile violation alerts issued for this account during the given time period, potentially filtered by security profile, behavior violated, or thing (device) violating.

listViolationEventsResponse_nextToken :: Lens' ListViolationEventsResponse (Maybe Text)

A token that can be used to retrieve the next set of results, or null if there are no additional results.

listViolationEventsResponse_httpStatus :: Lens' ListViolationEventsResponse Int

The response's http status code.