Amazonka.IoT.PutVerificationStateOnViolation

Description

Set a verification state and provide a description of that verification state on a violation (detect alarm).

Creating a Request

data PutVerificationStateOnViolation

See: newPutVerificationStateOnViolation smart constructor.

Constructors

PutVerificationStateOnViolation'

Fields

verificationStateDescription :: Maybe Text

The description of the verification state of the violation (detect alarm).

violationId :: Text

The violation ID.

verificationState :: VerificationState

The verification state of the violation.

newPutVerificationStateOnViolation

Arguments

:: Text
$sel:violationId:PutVerificationStateOnViolation'

-> VerificationState
$sel:verificationState:PutVerificationStateOnViolation'

-> PutVerificationStateOnViolation

Create a value of PutVerificationStateOnViolation with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:verificationStateDescription:PutVerificationStateOnViolation', putVerificationStateOnViolation_verificationStateDescription - The description of the verification state of the violation (detect alarm).

$sel:violationId:PutVerificationStateOnViolation', putVerificationStateOnViolation_violationId - The violation ID.

$sel:verificationState:PutVerificationStateOnViolation', putVerificationStateOnViolation_verificationState - The verification state of the violation.

Request Lenses

putVerificationStateOnViolation_verificationStateDescription :: Lens' PutVerificationStateOnViolation (Maybe Text)

The description of the verification state of the violation (detect alarm).

putVerificationStateOnViolation_violationId :: Lens' PutVerificationStateOnViolation Text

The violation ID.

putVerificationStateOnViolation_verificationState :: Lens' PutVerificationStateOnViolation VerificationState

The verification state of the violation.

Destructuring the Response

data PutVerificationStateOnViolationResponse

See: newPutVerificationStateOnViolationResponse smart constructor.

Constructors

PutVerificationStateOnViolationResponse'

Fields

httpStatus :: Int

The response's http status code.

newPutVerificationStateOnViolationResponse

Arguments

:: Int
$sel:httpStatus:PutVerificationStateOnViolationResponse'

-> PutVerificationStateOnViolationResponse

Create a value of PutVerificationStateOnViolationResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:httpStatus:PutVerificationStateOnViolationResponse', putVerificationStateOnViolationResponse_httpStatus - The response's http status code.

Response Lenses

putVerificationStateOnViolationResponse_httpStatus :: Lens' PutVerificationStateOnViolationResponse Int

The response's http status code.