Amazonka.IoT.StartAuditMitigationActionsTask

Description

Starts a task that applies a set of mitigation actions to the specified target.

Requires permission to access the StartAuditMitigationActionsTask action.

Creating a Request

data StartAuditMitigationActionsTask

See: newStartAuditMitigationActionsTask smart constructor.

Constructors

StartAuditMitigationActionsTask'

Fields

taskId :: Text

A unique identifier for the task. You can use this identifier to check the status of the task or to cancel it.

target :: AuditMitigationActionsTaskTarget

Specifies the audit findings to which the mitigation actions are applied. You can apply them to a type of audit check, to all findings from an audit, or to a specific set of findings.

auditCheckToActionsMapping :: HashMap Text (NonEmpty Text)

For an audit check, specifies which mitigation actions to apply. Those actions must be defined in your Amazon Web Services accounts.

clientRequestToken :: Text

Each audit mitigation task must have a unique client request token. If you try to start a new task with the same token as a task that already exists, an exception occurs. If you omit this value, a unique client request token is generated automatically.

newStartAuditMitigationActionsTask

Arguments

:: Text
$sel:taskId:StartAuditMitigationActionsTask'

-> AuditMitigationActionsTaskTarget
$sel:target:StartAuditMitigationActionsTask'

-> Text
$sel:clientRequestToken:StartAuditMitigationActionsTask'

-> StartAuditMitigationActionsTask

Create a value of StartAuditMitigationActionsTask with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:taskId:StartAuditMitigationActionsTask', startAuditMitigationActionsTask_taskId - A unique identifier for the task. You can use this identifier to check the status of the task or to cancel it.

$sel:target:StartAuditMitigationActionsTask', startAuditMitigationActionsTask_target - Specifies the audit findings to which the mitigation actions are applied. You can apply them to a type of audit check, to all findings from an audit, or to a specific set of findings.

$sel:auditCheckToActionsMapping:StartAuditMitigationActionsTask', startAuditMitigationActionsTask_auditCheckToActionsMapping - For an audit check, specifies which mitigation actions to apply. Those actions must be defined in your Amazon Web Services accounts.

$sel:clientRequestToken:StartAuditMitigationActionsTask', startAuditMitigationActionsTask_clientRequestToken - Each audit mitigation task must have a unique client request token. If you try to start a new task with the same token as a task that already exists, an exception occurs. If you omit this value, a unique client request token is generated automatically.

Request Lenses

startAuditMitigationActionsTask_taskId :: Lens' StartAuditMitigationActionsTask Text

A unique identifier for the task. You can use this identifier to check the status of the task or to cancel it.

startAuditMitigationActionsTask_target :: Lens' StartAuditMitigationActionsTask AuditMitigationActionsTaskTarget

Specifies the audit findings to which the mitigation actions are applied. You can apply them to a type of audit check, to all findings from an audit, or to a specific set of findings.

startAuditMitigationActionsTask_auditCheckToActionsMapping :: Lens' StartAuditMitigationActionsTask (HashMap Text (NonEmpty Text))

For an audit check, specifies which mitigation actions to apply. Those actions must be defined in your Amazon Web Services accounts.

startAuditMitigationActionsTask_clientRequestToken :: Lens' StartAuditMitigationActionsTask Text

Each audit mitigation task must have a unique client request token. If you try to start a new task with the same token as a task that already exists, an exception occurs. If you omit this value, a unique client request token is generated automatically.

Destructuring the Response

data StartAuditMitigationActionsTaskResponse

See: newStartAuditMitigationActionsTaskResponse smart constructor.

Constructors

StartAuditMitigationActionsTaskResponse'

Fields

taskId :: Maybe Text

The unique identifier for the audit mitigation task. This matches the taskId that you specified in the request.

httpStatus :: Int

The response's http status code.

newStartAuditMitigationActionsTaskResponse

Arguments

:: Int
$sel:httpStatus:StartAuditMitigationActionsTaskResponse'

-> StartAuditMitigationActionsTaskResponse

Create a value of StartAuditMitigationActionsTaskResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:taskId:StartAuditMitigationActionsTask', startAuditMitigationActionsTaskResponse_taskId - The unique identifier for the audit mitigation task. This matches the taskId that you specified in the request.

$sel:httpStatus:StartAuditMitigationActionsTaskResponse', startAuditMitigationActionsTaskResponse_httpStatus - The response's http status code.

Response Lenses

startAuditMitigationActionsTaskResponse_taskId :: Lens' StartAuditMitigationActionsTaskResponse (Maybe Text)

The unique identifier for the audit mitigation task. This matches the taskId that you specified in the request.

startAuditMitigationActionsTaskResponse_httpStatus :: Lens' StartAuditMitigationActionsTaskResponse Int

The response's http status code.