Amazonka.IoT.StartDetectMitigationActionsTask

Description

Starts a Device Defender ML Detect mitigation actions task.

Requires permission to access the StartDetectMitigationActionsTask action.

Creating a Request

data StartDetectMitigationActionsTask

See: newStartDetectMitigationActionsTask smart constructor.

Constructors

StartDetectMitigationActionsTask'

Fields

violationEventOccurrenceRange :: Maybe ViolationEventOccurrenceRange

Specifies the time period of which violation events occurred between.

includeOnlyActiveViolations :: Maybe Bool

Specifies to list only active violations.

includeSuppressedAlerts :: Maybe Bool

Specifies to include suppressed alerts.

taskId :: Text

The unique identifier of the task.

target :: DetectMitigationActionsTaskTarget

Specifies the ML Detect findings to which the mitigation actions are applied.

actions :: NonEmpty Text

The actions to be performed when a device has unexpected behavior.

clientRequestToken :: Text

Each mitigation action task must have a unique client request token. If you try to create a new task with the same token as a task that already exists, an exception occurs. If you omit this value, Amazon Web Services SDKs will automatically generate a unique client request.

newStartDetectMitigationActionsTask

Arguments

:: Text
$sel:taskId:StartDetectMitigationActionsTask'

-> DetectMitigationActionsTaskTarget
$sel:target:StartDetectMitigationActionsTask'

-> NonEmpty Text
$sel:actions:StartDetectMitigationActionsTask'

-> Text
$sel:clientRequestToken:StartDetectMitigationActionsTask'

-> StartDetectMitigationActionsTask

Create a value of StartDetectMitigationActionsTask with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:violationEventOccurrenceRange:StartDetectMitigationActionsTask', startDetectMitigationActionsTask_violationEventOccurrenceRange - Specifies the time period of which violation events occurred between.

$sel:includeOnlyActiveViolations:StartDetectMitigationActionsTask', startDetectMitigationActionsTask_includeOnlyActiveViolations - Specifies to list only active violations.

$sel:includeSuppressedAlerts:StartDetectMitigationActionsTask', startDetectMitigationActionsTask_includeSuppressedAlerts - Specifies to include suppressed alerts.

$sel:taskId:StartDetectMitigationActionsTask', startDetectMitigationActionsTask_taskId - The unique identifier of the task.

$sel:target:StartDetectMitigationActionsTask', startDetectMitigationActionsTask_target - Specifies the ML Detect findings to which the mitigation actions are applied.

$sel:actions:StartDetectMitigationActionsTask', startDetectMitigationActionsTask_actions - The actions to be performed when a device has unexpected behavior.

$sel:clientRequestToken:StartDetectMitigationActionsTask', startDetectMitigationActionsTask_clientRequestToken - Each mitigation action task must have a unique client request token. If you try to create a new task with the same token as a task that already exists, an exception occurs. If you omit this value, Amazon Web Services SDKs will automatically generate a unique client request.

Request Lenses

startDetectMitigationActionsTask_violationEventOccurrenceRange :: Lens' StartDetectMitigationActionsTask (Maybe ViolationEventOccurrenceRange)

Specifies the time period of which violation events occurred between.

startDetectMitigationActionsTask_includeOnlyActiveViolations :: Lens' StartDetectMitigationActionsTask (Maybe Bool)

Specifies to list only active violations.

startDetectMitigationActionsTask_includeSuppressedAlerts :: Lens' StartDetectMitigationActionsTask (Maybe Bool)

Specifies to include suppressed alerts.

startDetectMitigationActionsTask_taskId :: Lens' StartDetectMitigationActionsTask Text

The unique identifier of the task.

startDetectMitigationActionsTask_target :: Lens' StartDetectMitigationActionsTask DetectMitigationActionsTaskTarget

Specifies the ML Detect findings to which the mitigation actions are applied.

startDetectMitigationActionsTask_actions :: Lens' StartDetectMitigationActionsTask (NonEmpty Text)

The actions to be performed when a device has unexpected behavior.

startDetectMitigationActionsTask_clientRequestToken :: Lens' StartDetectMitigationActionsTask Text

Each mitigation action task must have a unique client request token. If you try to create a new task with the same token as a task that already exists, an exception occurs. If you omit this value, Amazon Web Services SDKs will automatically generate a unique client request.

Destructuring the Response

data StartDetectMitigationActionsTaskResponse

See: newStartDetectMitigationActionsTaskResponse smart constructor.

Constructors

StartDetectMitigationActionsTaskResponse'

Fields

taskId :: Maybe Text

The unique identifier of the task.

httpStatus :: Int

The response's http status code.

newStartDetectMitigationActionsTaskResponse

Arguments

:: Int
$sel:httpStatus:StartDetectMitigationActionsTaskResponse'

-> StartDetectMitigationActionsTaskResponse

Create a value of StartDetectMitigationActionsTaskResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:taskId:StartDetectMitigationActionsTask', startDetectMitigationActionsTaskResponse_taskId - The unique identifier of the task.

$sel:httpStatus:StartDetectMitigationActionsTaskResponse', startDetectMitigationActionsTaskResponse_httpStatus - The response's http status code.

Response Lenses

startDetectMitigationActionsTaskResponse_taskId :: Lens' StartDetectMitigationActionsTaskResponse (Maybe Text)

The unique identifier of the task.

startDetectMitigationActionsTaskResponse_httpStatus :: Lens' StartDetectMitigationActionsTaskResponse Int

The response's http status code.