Amazonka.IoT.TestInvokeAuthorizer

Description

Tests a custom authorization behavior by invoking a specified custom authorizer. Use this to test and debug the custom authorization behavior of devices that connect to the IoT device gateway.

Requires permission to access the TestInvokeAuthorizer action.

Creating a Request

data TestInvokeAuthorizer

See: newTestInvokeAuthorizer smart constructor.

Constructors

TestInvokeAuthorizer'

Fields

token :: Maybe Text

The token returned by your custom authentication service.

tlsContext :: Maybe TlsContext

Specifies a test TLS authorization request.

tokenSignature :: Maybe Text

The signature made with the token and your custom authentication service's private key. This value must be Base-64-encoded.

httpContext :: Maybe HttpContext

Specifies a test HTTP authorization request.

mqttContext :: Maybe MqttContext

Specifies a test MQTT authorization request.

authorizerName :: Text

The custom authorizer name.

newTestInvokeAuthorizer

Arguments

:: Text
$sel:authorizerName:TestInvokeAuthorizer'

-> TestInvokeAuthorizer

Create a value of TestInvokeAuthorizer with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:token:TestInvokeAuthorizer', testInvokeAuthorizer_token - The token returned by your custom authentication service.

$sel:tlsContext:TestInvokeAuthorizer', testInvokeAuthorizer_tlsContext - Specifies a test TLS authorization request.

$sel:tokenSignature:TestInvokeAuthorizer', testInvokeAuthorizer_tokenSignature - The signature made with the token and your custom authentication service's private key. This value must be Base-64-encoded.

$sel:httpContext:TestInvokeAuthorizer', testInvokeAuthorizer_httpContext - Specifies a test HTTP authorization request.

$sel:mqttContext:TestInvokeAuthorizer', testInvokeAuthorizer_mqttContext - Specifies a test MQTT authorization request.

$sel:authorizerName:TestInvokeAuthorizer', testInvokeAuthorizer_authorizerName - The custom authorizer name.

Request Lenses

testInvokeAuthorizer_token :: Lens' TestInvokeAuthorizer (Maybe Text)

The token returned by your custom authentication service.

testInvokeAuthorizer_tlsContext :: Lens' TestInvokeAuthorizer (Maybe TlsContext)

Specifies a test TLS authorization request.

testInvokeAuthorizer_tokenSignature :: Lens' TestInvokeAuthorizer (Maybe Text)

The signature made with the token and your custom authentication service's private key. This value must be Base-64-encoded.

testInvokeAuthorizer_httpContext :: Lens' TestInvokeAuthorizer (Maybe HttpContext)

Specifies a test HTTP authorization request.

testInvokeAuthorizer_mqttContext :: Lens' TestInvokeAuthorizer (Maybe MqttContext)

Specifies a test MQTT authorization request.

testInvokeAuthorizer_authorizerName :: Lens' TestInvokeAuthorizer Text

The custom authorizer name.

Destructuring the Response

data TestInvokeAuthorizerResponse

See: newTestInvokeAuthorizerResponse smart constructor.

Constructors

TestInvokeAuthorizerResponse'

Fields

policyDocuments :: Maybe [Text]

IAM policy documents.

principalId :: Maybe Text

The principal ID.

disconnectAfterInSeconds :: Maybe Int

The number of seconds after which the connection is terminated.

isAuthenticated :: Maybe Bool

True if the token is authenticated, otherwise false.

refreshAfterInSeconds :: Maybe Int

The number of seconds after which the temporary credentials are refreshed.

httpStatus :: Int

The response's http status code.

newTestInvokeAuthorizerResponse

Arguments

:: Int
$sel:httpStatus:TestInvokeAuthorizerResponse'

-> TestInvokeAuthorizerResponse

Create a value of TestInvokeAuthorizerResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:policyDocuments:TestInvokeAuthorizerResponse', testInvokeAuthorizerResponse_policyDocuments - IAM policy documents.

$sel:principalId:TestInvokeAuthorizerResponse', testInvokeAuthorizerResponse_principalId - The principal ID.

$sel:disconnectAfterInSeconds:TestInvokeAuthorizerResponse', testInvokeAuthorizerResponse_disconnectAfterInSeconds - The number of seconds after which the connection is terminated.

$sel:isAuthenticated:TestInvokeAuthorizerResponse', testInvokeAuthorizerResponse_isAuthenticated - True if the token is authenticated, otherwise false.

$sel:refreshAfterInSeconds:TestInvokeAuthorizerResponse', testInvokeAuthorizerResponse_refreshAfterInSeconds - The number of seconds after which the temporary credentials are refreshed.

$sel:httpStatus:TestInvokeAuthorizerResponse', testInvokeAuthorizerResponse_httpStatus - The response's http status code.

Response Lenses

testInvokeAuthorizerResponse_policyDocuments :: Lens' TestInvokeAuthorizerResponse (Maybe [Text])

IAM policy documents.

testInvokeAuthorizerResponse_principalId :: Lens' TestInvokeAuthorizerResponse (Maybe Text)

The principal ID.

testInvokeAuthorizerResponse_disconnectAfterInSeconds :: Lens' TestInvokeAuthorizerResponse (Maybe Int)

The number of seconds after which the connection is terminated.

testInvokeAuthorizerResponse_isAuthenticated :: Lens' TestInvokeAuthorizerResponse (Maybe Bool)

True if the token is authenticated, otherwise false.

testInvokeAuthorizerResponse_refreshAfterInSeconds :: Lens' TestInvokeAuthorizerResponse (Maybe Int)

The number of seconds after which the temporary credentials are refreshed.

testInvokeAuthorizerResponse_httpStatus :: Lens' TestInvokeAuthorizerResponse Int

The response's http status code.