Amazonka.IoT.TransferCertificate

Description

Transfers the specified certificate to the specified Amazon Web Services account.

Requires permission to access the TransferCertificate action.

You can cancel the transfer until it is acknowledged by the recipient.

No notification is sent to the transfer destination's account. It is up to the caller to notify the transfer target.

The certificate being transferred must not be in the ACTIVE state. You can use the UpdateCertificate action to deactivate it.

The certificate must not have any policies attached to it. You can use the DetachPolicy action to detach them.

Creating a Request

data TransferCertificate

The input for the TransferCertificate operation.

See: newTransferCertificate smart constructor.

Constructors

TransferCertificate'

Fields

transferMessage :: Maybe Text

The transfer message.

certificateId :: Text

The ID of the certificate. (The last part of the certificate ARN contains the certificate ID.)

targetAwsAccount :: Text

The Amazon Web Services account.

newTransferCertificate

Arguments

:: Text
$sel:certificateId:TransferCertificate'

-> Text
$sel:targetAwsAccount:TransferCertificate'

-> TransferCertificate

Create a value of TransferCertificate with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:transferMessage:TransferCertificate', transferCertificate_transferMessage - The transfer message.

$sel:certificateId:TransferCertificate', transferCertificate_certificateId - The ID of the certificate. (The last part of the certificate ARN contains the certificate ID.)

$sel:targetAwsAccount:TransferCertificate', transferCertificate_targetAwsAccount - The Amazon Web Services account.

Request Lenses

transferCertificate_transferMessage :: Lens' TransferCertificate (Maybe Text)

The transfer message.

transferCertificate_certificateId :: Lens' TransferCertificate Text

The ID of the certificate. (The last part of the certificate ARN contains the certificate ID.)

transferCertificate_targetAwsAccount :: Lens' TransferCertificate Text

The Amazon Web Services account.

Destructuring the Response

data TransferCertificateResponse

The output from the TransferCertificate operation.

See: newTransferCertificateResponse smart constructor.

Constructors

TransferCertificateResponse'

Fields

transferredCertificateArn :: Maybe Text

The ARN of the certificate.

httpStatus :: Int

The response's http status code.

newTransferCertificateResponse

Arguments

:: Int
$sel:httpStatus:TransferCertificateResponse'

-> TransferCertificateResponse

Create a value of TransferCertificateResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:transferredCertificateArn:TransferCertificateResponse', transferCertificateResponse_transferredCertificateArn - The ARN of the certificate.

$sel:httpStatus:TransferCertificateResponse', transferCertificateResponse_httpStatus - The response's http status code.

Response Lenses

transferCertificateResponse_transferredCertificateArn :: Lens' TransferCertificateResponse (Maybe Text)

The ARN of the certificate.

transferCertificateResponse_httpStatus :: Lens' TransferCertificateResponse Int

The response's http status code.