Amazonka.IoT.Types.AssetPropertyVariant

Description

Documentation

data AssetPropertyVariant

Contains an asset property value (of a single type).

See: newAssetPropertyVariant smart constructor.

Constructors

AssetPropertyVariant'

Fields

integerValue :: Maybe Text

Optional. A string that contains the integer value of the value entry. Accepts substitution templates.

doubleValue :: Maybe Text

Optional. A string that contains the double value of the value entry. Accepts substitution templates.

stringValue :: Maybe Text

Optional. The string value of the value entry. Accepts substitution templates.

booleanValue :: Maybe Text

Optional. A string that contains the boolean value (true or false) of the value entry. Accepts substitution templates.

newAssetPropertyVariant :: AssetPropertyVariant

Create a value of AssetPropertyVariant with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:integerValue:AssetPropertyVariant', assetPropertyVariant_integerValue - Optional. A string that contains the integer value of the value entry. Accepts substitution templates.

$sel:doubleValue:AssetPropertyVariant', assetPropertyVariant_doubleValue - Optional. A string that contains the double value of the value entry. Accepts substitution templates.

$sel:stringValue:AssetPropertyVariant', assetPropertyVariant_stringValue - Optional. The string value of the value entry. Accepts substitution templates.

$sel:booleanValue:AssetPropertyVariant', assetPropertyVariant_booleanValue - Optional. A string that contains the boolean value (true or false) of the value entry. Accepts substitution templates.

assetPropertyVariant_integerValue :: Lens' AssetPropertyVariant (Maybe Text)

Optional. A string that contains the integer value of the value entry. Accepts substitution templates.

assetPropertyVariant_doubleValue :: Lens' AssetPropertyVariant (Maybe Text)

Optional. A string that contains the double value of the value entry. Accepts substitution templates.

assetPropertyVariant_stringValue :: Lens' AssetPropertyVariant (Maybe Text)

Optional. The string value of the value entry. Accepts substitution templates.

assetPropertyVariant_booleanValue :: Lens' AssetPropertyVariant (Maybe Text)

Optional. A string that contains the boolean value (true or false) of the value entry. Accepts substitution templates.