Amazonka.IoT.Types.AuditFinding

Description

Documentation

data AuditFinding

The findings (results) of the audit.

See: newAuditFinding smart constructor.

Constructors

AuditFinding'

Fields

isSuppressed :: Maybe Bool

Indicates whether the audit finding was suppressed or not during reporting.

taskId :: Maybe Text

The ID of the audit that generated this result (finding).

findingTime :: Maybe POSIX

The time the result (finding) was discovered.

taskStartTime :: Maybe POSIX

The time the audit started.

reasonForNonComplianceCode :: Maybe Text

A code that indicates the reason that the resource was noncompliant.

severity :: Maybe AuditFindingSeverity

The severity of the result (finding).

relatedResources :: Maybe [RelatedResource]

The list of related resources.

checkName :: Maybe Text

The audit check that generated this result.

nonCompliantResource :: Maybe NonCompliantResource

The resource that was found to be noncompliant with the audit check.

reasonForNonCompliance :: Maybe Text

The reason the resource was noncompliant.

findingId :: Maybe Text

A unique identifier for this set of audit findings. This identifier is used to apply mitigation tasks to one or more sets of findings.

newAuditFinding :: AuditFinding

Create a value of AuditFinding with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:isSuppressed:AuditFinding', auditFinding_isSuppressed - Indicates whether the audit finding was suppressed or not during reporting.

$sel:taskId:AuditFinding', auditFinding_taskId - The ID of the audit that generated this result (finding).

$sel:findingTime:AuditFinding', auditFinding_findingTime - The time the result (finding) was discovered.

$sel:taskStartTime:AuditFinding', auditFinding_taskStartTime - The time the audit started.

$sel:reasonForNonComplianceCode:AuditFinding', auditFinding_reasonForNonComplianceCode - A code that indicates the reason that the resource was noncompliant.

$sel:severity:AuditFinding', auditFinding_severity - The severity of the result (finding).

$sel:relatedResources:AuditFinding', auditFinding_relatedResources - The list of related resources.

$sel:checkName:AuditFinding', auditFinding_checkName - The audit check that generated this result.

$sel:nonCompliantResource:AuditFinding', auditFinding_nonCompliantResource - The resource that was found to be noncompliant with the audit check.

$sel:reasonForNonCompliance:AuditFinding', auditFinding_reasonForNonCompliance - The reason the resource was noncompliant.

$sel:findingId:AuditFinding', auditFinding_findingId - A unique identifier for this set of audit findings. This identifier is used to apply mitigation tasks to one or more sets of findings.

auditFinding_isSuppressed :: Lens' AuditFinding (Maybe Bool)

Indicates whether the audit finding was suppressed or not during reporting.

auditFinding_taskId :: Lens' AuditFinding (Maybe Text)

The ID of the audit that generated this result (finding).

auditFinding_findingTime :: Lens' AuditFinding (Maybe UTCTime)

The time the result (finding) was discovered.

auditFinding_taskStartTime :: Lens' AuditFinding (Maybe UTCTime)

The time the audit started.

auditFinding_reasonForNonComplianceCode :: Lens' AuditFinding (Maybe Text)

A code that indicates the reason that the resource was noncompliant.

auditFinding_severity :: Lens' AuditFinding (Maybe AuditFindingSeverity)

The severity of the result (finding).

auditFinding_relatedResources :: Lens' AuditFinding (Maybe [RelatedResource])

The list of related resources.

auditFinding_checkName :: Lens' AuditFinding (Maybe Text)

The audit check that generated this result.

auditFinding_nonCompliantResource :: Lens' AuditFinding (Maybe NonCompliantResource)

The resource that was found to be noncompliant with the audit check.

auditFinding_reasonForNonCompliance :: Lens' AuditFinding (Maybe Text)

The reason the resource was noncompliant.

auditFinding_findingId :: Lens' AuditFinding (Maybe Text)

A unique identifier for this set of audit findings. This identifier is used to apply mitigation tasks to one or more sets of findings.