Amazonka.IoT.Types.DetectMitigationActionsTaskSummary

Description

Documentation

data DetectMitigationActionsTaskSummary

The summary of the mitigation action tasks.

See: newDetectMitigationActionsTaskSummary smart constructor.

Constructors

DetectMitigationActionsTaskSummary'

Fields

onlyActiveViolationsIncluded :: Maybe Bool

Includes only active violations.

suppressedAlertsIncluded :: Maybe Bool

Includes suppressed alerts.

violationEventOccurrenceRange :: Maybe ViolationEventOccurrenceRange

Specifies the time period of which violation events occurred between.

taskId :: Maybe Text

The unique identifier of the task.

taskStartTime :: Maybe POSIX

The date the task started.

taskStatistics :: Maybe DetectMitigationActionsTaskStatistics

The statistics of a mitigation action task.

actionsDefinition :: Maybe [MitigationAction]

The definition of the actions.

taskEndTime :: Maybe POSIX

The date the task ended.

target :: Maybe DetectMitigationActionsTaskTarget

Specifies the ML Detect findings to which the mitigation actions are applied.

taskStatus :: Maybe DetectMitigationActionsTaskStatus

The status of the task.

newDetectMitigationActionsTaskSummary :: DetectMitigationActionsTaskSummary

Create a value of DetectMitigationActionsTaskSummary with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:onlyActiveViolationsIncluded:DetectMitigationActionsTaskSummary', detectMitigationActionsTaskSummary_onlyActiveViolationsIncluded - Includes only active violations.

$sel:suppressedAlertsIncluded:DetectMitigationActionsTaskSummary', detectMitigationActionsTaskSummary_suppressedAlertsIncluded - Includes suppressed alerts.

$sel:violationEventOccurrenceRange:DetectMitigationActionsTaskSummary', detectMitigationActionsTaskSummary_violationEventOccurrenceRange - Specifies the time period of which violation events occurred between.

$sel:taskId:DetectMitigationActionsTaskSummary', detectMitigationActionsTaskSummary_taskId - The unique identifier of the task.

$sel:taskStartTime:DetectMitigationActionsTaskSummary', detectMitigationActionsTaskSummary_taskStartTime - The date the task started.

$sel:taskStatistics:DetectMitigationActionsTaskSummary', detectMitigationActionsTaskSummary_taskStatistics - The statistics of a mitigation action task.

$sel:actionsDefinition:DetectMitigationActionsTaskSummary', detectMitigationActionsTaskSummary_actionsDefinition - The definition of the actions.

$sel:taskEndTime:DetectMitigationActionsTaskSummary', detectMitigationActionsTaskSummary_taskEndTime - The date the task ended.

$sel:target:DetectMitigationActionsTaskSummary', detectMitigationActionsTaskSummary_target - Specifies the ML Detect findings to which the mitigation actions are applied.

$sel:taskStatus:DetectMitigationActionsTaskSummary', detectMitigationActionsTaskSummary_taskStatus - The status of the task.

detectMitigationActionsTaskSummary_onlyActiveViolationsIncluded :: Lens' DetectMitigationActionsTaskSummary (Maybe Bool)

Includes only active violations.

detectMitigationActionsTaskSummary_suppressedAlertsIncluded :: Lens' DetectMitigationActionsTaskSummary (Maybe Bool)

Includes suppressed alerts.

detectMitigationActionsTaskSummary_violationEventOccurrenceRange :: Lens' DetectMitigationActionsTaskSummary (Maybe ViolationEventOccurrenceRange)

Specifies the time period of which violation events occurred between.

detectMitigationActionsTaskSummary_taskId :: Lens' DetectMitigationActionsTaskSummary (Maybe Text)

The unique identifier of the task.

detectMitigationActionsTaskSummary_taskStartTime :: Lens' DetectMitigationActionsTaskSummary (Maybe UTCTime)

The date the task started.

detectMitigationActionsTaskSummary_taskStatistics :: Lens' DetectMitigationActionsTaskSummary (Maybe DetectMitigationActionsTaskStatistics)

The statistics of a mitigation action task.

detectMitigationActionsTaskSummary_actionsDefinition :: Lens' DetectMitigationActionsTaskSummary (Maybe [MitigationAction])

The definition of the actions.

detectMitigationActionsTaskSummary_taskEndTime :: Lens' DetectMitigationActionsTaskSummary (Maybe UTCTime)

The date the task ended.

detectMitigationActionsTaskSummary_target :: Lens' DetectMitigationActionsTaskSummary (Maybe DetectMitigationActionsTaskTarget)

Specifies the ML Detect findings to which the mitigation actions are applied.

detectMitigationActionsTaskSummary_taskStatus :: Lens' DetectMitigationActionsTaskSummary (Maybe DetectMitigationActionsTaskStatus)

The status of the task.