Amazonka.IoT.Types.EnableIoTLoggingParams

Description

Documentation

data EnableIoTLoggingParams

Parameters used when defining a mitigation action that enable Amazon Web Services IoT Core logging.

See: newEnableIoTLoggingParams smart constructor.

Constructors

EnableIoTLoggingParams'

Fields

roleArnForLogging :: Text

The Amazon Resource Name (ARN) of the IAM role used for logging.

logLevel :: LogLevel

Specifies the type of information to be logged.

newEnableIoTLoggingParams

Arguments

:: Text
$sel:roleArnForLogging:EnableIoTLoggingParams'

-> LogLevel
$sel:logLevel:EnableIoTLoggingParams'

-> EnableIoTLoggingParams

Create a value of EnableIoTLoggingParams with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:roleArnForLogging:EnableIoTLoggingParams', enableIoTLoggingParams_roleArnForLogging - The Amazon Resource Name (ARN) of the IAM role used for logging.

$sel:logLevel:EnableIoTLoggingParams', enableIoTLoggingParams_logLevel - Specifies the type of information to be logged.

enableIoTLoggingParams_roleArnForLogging :: Lens' EnableIoTLoggingParams Text

The Amazon Resource Name (ARN) of the IAM role used for logging.

enableIoTLoggingParams_logLevel :: Lens' EnableIoTLoggingParams LogLevel

Specifies the type of information to be logged.