Amazonka.IoT.Types.JobExecutionSummaryForJob

Description

Documentation

data JobExecutionSummaryForJob

Contains a summary of information about job executions for a specific job.

See: newJobExecutionSummaryForJob smart constructor.

Constructors

JobExecutionSummaryForJob'

Fields

jobExecutionSummary :: Maybe JobExecutionSummary

Contains a subset of information about a job execution.

thingArn :: Maybe Text

The ARN of the thing on which the job execution is running.

newJobExecutionSummaryForJob :: JobExecutionSummaryForJob

Create a value of JobExecutionSummaryForJob with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:jobExecutionSummary:JobExecutionSummaryForJob', jobExecutionSummaryForJob_jobExecutionSummary - Contains a subset of information about a job execution.

$sel:thingArn:JobExecutionSummaryForJob', jobExecutionSummaryForJob_thingArn - The ARN of the thing on which the job execution is running.

jobExecutionSummaryForJob_jobExecutionSummary :: Lens' JobExecutionSummaryForJob (Maybe JobExecutionSummary)

Contains a subset of information about a job execution.

jobExecutionSummaryForJob_thingArn :: Lens' JobExecutionSummaryForJob (Maybe Text)

The ARN of the thing on which the job execution is running.