Amazonka.IoT.Types.KafkaAction

Description

Documentation

data KafkaAction

Send messages to an Amazon Managed Streaming for Apache Kafka (Amazon MSK) or self-managed Apache Kafka cluster.

See: newKafkaAction smart constructor.

Constructors

KafkaAction'

Fields

key :: Maybe Text

The Kafka message key.

partition :: Maybe Text

The Kafka message partition.

destinationArn :: Text

The ARN of Kafka action's VPC TopicRuleDestination.

topic :: Text

The Kafka topic for messages to be sent to the Kafka broker.

clientProperties :: HashMap Text Text

Properties of the Apache Kafka producer client.

newKafkaAction

Arguments

:: Text
$sel:destinationArn:KafkaAction'

-> Text
$sel:topic:KafkaAction'

-> KafkaAction

Create a value of KafkaAction with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:key:KafkaAction', kafkaAction_key - The Kafka message key.

$sel:partition:KafkaAction', kafkaAction_partition - The Kafka message partition.

$sel:destinationArn:KafkaAction', kafkaAction_destinationArn - The ARN of Kafka action's VPC TopicRuleDestination.

$sel:topic:KafkaAction', kafkaAction_topic - The Kafka topic for messages to be sent to the Kafka broker.

$sel:clientProperties:KafkaAction', kafkaAction_clientProperties - Properties of the Apache Kafka producer client.

kafkaAction_key :: Lens' KafkaAction (Maybe Text)

The Kafka message key.

kafkaAction_partition :: Lens' KafkaAction (Maybe Text)

The Kafka message partition.

kafkaAction_destinationArn :: Lens' KafkaAction Text

The ARN of Kafka action's VPC TopicRuleDestination.

kafkaAction_topic :: Lens' KafkaAction Text

The Kafka topic for messages to be sent to the Kafka broker.

kafkaAction_clientProperties :: Lens' KafkaAction (HashMap Text Text)

Properties of the Apache Kafka producer client.