Amazonka.IoT.Types.OpenSearchAction

Description

Documentation

data OpenSearchAction

Describes an action that writes data to an Amazon OpenSearch Service domain.

See: newOpenSearchAction smart constructor.

Constructors

OpenSearchAction'

Fields

roleArn :: Text

The IAM role ARN that has access to OpenSearch.

endpoint :: Text

The endpoint of your OpenSearch domain.

index :: Text

The OpenSearch index where you want to store your data.

type' :: Text

The type of document you are storing.

id :: Text

The unique identifier for the document you are storing.

newOpenSearchAction

Arguments

:: Text
$sel:roleArn:OpenSearchAction'

-> Text
$sel:endpoint:OpenSearchAction'

-> Text
$sel:index:OpenSearchAction'

-> Text
$sel:type':OpenSearchAction'

-> Text
$sel:id:OpenSearchAction'

-> OpenSearchAction

Create a value of OpenSearchAction with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:roleArn:OpenSearchAction', openSearchAction_roleArn - The IAM role ARN that has access to OpenSearch.

$sel:endpoint:OpenSearchAction', openSearchAction_endpoint - The endpoint of your OpenSearch domain.

$sel:index:OpenSearchAction', openSearchAction_index - The OpenSearch index where you want to store your data.

$sel:type':OpenSearchAction', openSearchAction_type - The type of document you are storing.

$sel:id:OpenSearchAction', openSearchAction_id - The unique identifier for the document you are storing.

openSearchAction_roleArn :: Lens' OpenSearchAction Text

The IAM role ARN that has access to OpenSearch.

openSearchAction_endpoint :: Lens' OpenSearchAction Text

The endpoint of your OpenSearch domain.

openSearchAction_index :: Lens' OpenSearchAction Text

The OpenSearch index where you want to store your data.

openSearchAction_type :: Lens' OpenSearchAction Text

The type of document you are storing.

openSearchAction_id :: Lens' OpenSearchAction Text

The unique identifier for the document you are storing.