Amazonka.IoT.Types.SigV4Authorization

Description

Documentation

data SigV4Authorization

For more information, see Signature Version 4 signing process.

See: newSigV4Authorization smart constructor.

Constructors

SigV4Authorization'

Fields

signingRegion :: Text

The signing region.

serviceName :: Text

The service name to use while signing with Sig V4.

roleArn :: Text

The ARN of the signing role.

newSigV4Authorization

Arguments

:: Text
$sel:signingRegion:SigV4Authorization'

-> Text
$sel:serviceName:SigV4Authorization'

-> Text
$sel:roleArn:SigV4Authorization'

-> SigV4Authorization

Create a value of SigV4Authorization with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:signingRegion:SigV4Authorization', sigV4Authorization_signingRegion - The signing region.

$sel:serviceName:SigV4Authorization', sigV4Authorization_serviceName - The service name to use while signing with Sig V4.

$sel:roleArn:SigV4Authorization', sigV4Authorization_roleArn - The ARN of the signing role.

sigV4Authorization_signingRegion :: Lens' SigV4Authorization Text

The signing region.

sigV4Authorization_serviceName :: Lens' SigV4Authorization Text

The service name to use while signing with Sig V4.

sigV4Authorization_roleArn :: Lens' SigV4Authorization Text

The ARN of the signing role.