Amazonka.IoT.Types.TaskStatistics

Description

Documentation

data TaskStatistics

Statistics for the checks performed during the audit.

See: newTaskStatistics smart constructor.

Constructors

TaskStatistics'

Fields

nonCompliantChecks :: Maybe Int

The number of checks that found noncompliant resources.

waitingForDataCollectionChecks :: Maybe Int

The number of checks waiting for data collection.

failedChecks :: Maybe Int

The number of checks.

totalChecks :: Maybe Int

The number of checks in this audit.

inProgressChecks :: Maybe Int

The number of checks in progress.

compliantChecks :: Maybe Int

The number of checks that found compliant resources.

canceledChecks :: Maybe Int

The number of checks that did not run because the audit was canceled.

newTaskStatistics :: TaskStatistics

Create a value of TaskStatistics with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:nonCompliantChecks:TaskStatistics', taskStatistics_nonCompliantChecks - The number of checks that found noncompliant resources.

$sel:waitingForDataCollectionChecks:TaskStatistics', taskStatistics_waitingForDataCollectionChecks - The number of checks waiting for data collection.

$sel:failedChecks:TaskStatistics', taskStatistics_failedChecks - The number of checks.

$sel:totalChecks:TaskStatistics', taskStatistics_totalChecks - The number of checks in this audit.

$sel:inProgressChecks:TaskStatistics', taskStatistics_inProgressChecks - The number of checks in progress.

$sel:compliantChecks:TaskStatistics', taskStatistics_compliantChecks - The number of checks that found compliant resources.

$sel:canceledChecks:TaskStatistics', taskStatistics_canceledChecks - The number of checks that did not run because the audit was canceled.

taskStatistics_nonCompliantChecks :: Lens' TaskStatistics (Maybe Int)

The number of checks that found noncompliant resources.

taskStatistics_waitingForDataCollectionChecks :: Lens' TaskStatistics (Maybe Int)

The number of checks waiting for data collection.

taskStatistics_failedChecks :: Lens' TaskStatistics (Maybe Int)

The number of checks.

taskStatistics_totalChecks :: Lens' TaskStatistics (Maybe Int)

The number of checks in this audit.

taskStatistics_inProgressChecks :: Lens' TaskStatistics (Maybe Int)

The number of checks in progress.

taskStatistics_compliantChecks :: Lens' TaskStatistics (Maybe Int)

The number of checks that found compliant resources.

taskStatistics_canceledChecks :: Lens' TaskStatistics (Maybe Int)

The number of checks that did not run because the audit was canceled.