Amazonka.IoT.Types.TlsContext

Description

Documentation

data TlsContext

Specifies the TLS context to use for the test authorizer request.

See: newTlsContext smart constructor.

Constructors

TlsContext'

Fields

serverName :: Maybe Text

The value of the serverName key in a TLS authorization request.

newTlsContext :: TlsContext

Create a value of TlsContext with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:serverName:TlsContext', tlsContext_serverName - The value of the serverName key in a TLS authorization request.

tlsContext_serverName :: Lens' TlsContext (Maybe Text)

The value of the serverName key in a TLS authorization request.