Amazonka.IoT.Types.TransferData

Description

Documentation

data TransferData

Data used to transfer a certificate to an Amazon Web Services account.

See: newTransferData smart constructor.

Constructors

TransferData'

Fields

transferDate :: Maybe POSIX

The date the transfer took place.

acceptDate :: Maybe POSIX

The date the transfer was accepted.

transferMessage :: Maybe Text

The transfer message.

rejectDate :: Maybe POSIX

The date the transfer was rejected.

rejectReason :: Maybe Text

The reason why the transfer was rejected.

newTransferData :: TransferData

Create a value of TransferData with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:transferDate:TransferData', transferData_transferDate - The date the transfer took place.

$sel:acceptDate:TransferData', transferData_acceptDate - The date the transfer was accepted.

$sel:transferMessage:TransferData', transferData_transferMessage - The transfer message.

$sel:rejectDate:TransferData', transferData_rejectDate - The date the transfer was rejected.

$sel:rejectReason:TransferData', transferData_rejectReason - The reason why the transfer was rejected.

transferData_transferDate :: Lens' TransferData (Maybe UTCTime)

The date the transfer took place.

transferData_acceptDate :: Lens' TransferData (Maybe UTCTime)

The date the transfer was accepted.

transferData_transferMessage :: Lens' TransferData (Maybe Text)

The transfer message.

transferData_rejectDate :: Lens' TransferData (Maybe UTCTime)

The date the transfer was rejected.

transferData_rejectReason :: Lens' TransferData (Maybe Text)

The reason why the transfer was rejected.