Amazonka.IoT.Types.VpcDestinationConfiguration

Description

Documentation

data VpcDestinationConfiguration

The configuration information for a virtual private cloud (VPC) destination.

See: newVpcDestinationConfiguration smart constructor.

Constructors

VpcDestinationConfiguration'

Fields

securityGroups :: Maybe [Text]

The security groups of the VPC destination.

subnetIds :: [Text]

The subnet IDs of the VPC destination.

vpcId :: Text

The ID of the VPC.

roleArn :: Text

The ARN of a role that has permission to create and attach to elastic network interfaces (ENIs).

newVpcDestinationConfiguration

Arguments

:: Text
$sel:vpcId:VpcDestinationConfiguration'

-> Text
$sel:roleArn:VpcDestinationConfiguration'

-> VpcDestinationConfiguration

Create a value of VpcDestinationConfiguration with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:securityGroups:VpcDestinationConfiguration', vpcDestinationConfiguration_securityGroups - The security groups of the VPC destination.

$sel:subnetIds:VpcDestinationConfiguration', vpcDestinationConfiguration_subnetIds - The subnet IDs of the VPC destination.

$sel:vpcId:VpcDestinationConfiguration', vpcDestinationConfiguration_vpcId - The ID of the VPC.

$sel:roleArn:VpcDestinationConfiguration', vpcDestinationConfiguration_roleArn - The ARN of a role that has permission to create and attach to elastic network interfaces (ENIs).

vpcDestinationConfiguration_securityGroups :: Lens' VpcDestinationConfiguration (Maybe [Text])

The security groups of the VPC destination.

vpcDestinationConfiguration_subnetIds :: Lens' VpcDestinationConfiguration [Text]

The subnet IDs of the VPC destination.

vpcDestinationConfiguration_vpcId :: Lens' VpcDestinationConfiguration Text

The ID of the VPC.

vpcDestinationConfiguration_roleArn :: Lens' VpcDestinationConfiguration Text

The ARN of a role that has permission to create and attach to elastic network interfaces (ENIs).