Amazonka.IoT.Types.VpcDestinationProperties

Description

Documentation

data VpcDestinationProperties

The properties of a virtual private cloud (VPC) destination.

See: newVpcDestinationProperties smart constructor.

Constructors

VpcDestinationProperties'

Fields

securityGroups :: Maybe [Text]

The security groups of the VPC destination.

subnetIds :: Maybe [Text]

The subnet IDs of the VPC destination.

vpcId :: Maybe Text

The ID of the VPC.

roleArn :: Maybe Text

The ARN of a role that has permission to create and attach to elastic network interfaces (ENIs).

newVpcDestinationProperties :: VpcDestinationProperties

Create a value of VpcDestinationProperties with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:securityGroups:VpcDestinationProperties', vpcDestinationProperties_securityGroups - The security groups of the VPC destination.

$sel:subnetIds:VpcDestinationProperties', vpcDestinationProperties_subnetIds - The subnet IDs of the VPC destination.

$sel:vpcId:VpcDestinationProperties', vpcDestinationProperties_vpcId - The ID of the VPC.

$sel:roleArn:VpcDestinationProperties', vpcDestinationProperties_roleArn - The ARN of a role that has permission to create and attach to elastic network interfaces (ENIs).

vpcDestinationProperties_securityGroups :: Lens' VpcDestinationProperties (Maybe [Text])

The security groups of the VPC destination.

vpcDestinationProperties_subnetIds :: Lens' VpcDestinationProperties (Maybe [Text])

The subnet IDs of the VPC destination.

vpcDestinationProperties_vpcId :: Lens' VpcDestinationProperties (Maybe Text)

The ID of the VPC.

vpcDestinationProperties_roleArn :: Lens' VpcDestinationProperties (Maybe Text)

The ARN of a role that has permission to create and attach to elastic network interfaces (ENIs).