Amazonka.IoT.UpdateRoleAlias

Description

Updates a role alias.

Requires permission to access the UpdateRoleAlias action.

Creating a Request

data UpdateRoleAlias

See: newUpdateRoleAlias smart constructor.

Constructors

UpdateRoleAlias'

Fields

credentialDurationSeconds :: Maybe Natural

The number of seconds the credential will be valid.

roleArn :: Maybe Text

The role ARN.

roleAlias :: Text

The role alias to update.

newUpdateRoleAlias

Arguments

:: Text
$sel:roleAlias:UpdateRoleAlias'

-> UpdateRoleAlias

Create a value of UpdateRoleAlias with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:credentialDurationSeconds:UpdateRoleAlias', updateRoleAlias_credentialDurationSeconds - The number of seconds the credential will be valid.

$sel:roleArn:UpdateRoleAlias', updateRoleAlias_roleArn - The role ARN.

$sel:roleAlias:UpdateRoleAlias', updateRoleAlias_roleAlias - The role alias to update.

Request Lenses

updateRoleAlias_credentialDurationSeconds :: Lens' UpdateRoleAlias (Maybe Natural)

The number of seconds the credential will be valid.

updateRoleAlias_roleArn :: Lens' UpdateRoleAlias (Maybe Text)

The role ARN.

updateRoleAlias_roleAlias :: Lens' UpdateRoleAlias Text

The role alias to update.

Destructuring the Response

data UpdateRoleAliasResponse

See: newUpdateRoleAliasResponse smart constructor.

Constructors

UpdateRoleAliasResponse'

Fields

roleAliasArn :: Maybe Text

The role alias ARN.

roleAlias :: Maybe Text

The role alias.

httpStatus :: Int

The response's http status code.

newUpdateRoleAliasResponse

Arguments

:: Int
$sel:httpStatus:UpdateRoleAliasResponse'

-> UpdateRoleAliasResponse

Create a value of UpdateRoleAliasResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:roleAliasArn:UpdateRoleAliasResponse', updateRoleAliasResponse_roleAliasArn - The role alias ARN.

$sel:roleAlias:UpdateRoleAlias', updateRoleAliasResponse_roleAlias - The role alias.

$sel:httpStatus:UpdateRoleAliasResponse', updateRoleAliasResponse_httpStatus - The response's http status code.

Response Lenses

updateRoleAliasResponse_roleAliasArn :: Lens' UpdateRoleAliasResponse (Maybe Text)

The role alias ARN.

updateRoleAliasResponse_roleAlias :: Lens' UpdateRoleAliasResponse (Maybe Text)

The role alias.

updateRoleAliasResponse_httpStatus :: Lens' UpdateRoleAliasResponse Int

The response's http status code.