Amazonka.IoT1ClickDevices

Description

Derived from API version 2018-05-14 of the AWS service descriptions, licensed under Apache 2.0.

Describes all of the AWS IoT 1-Click device-related API operations for the service. Also provides sample requests, responses, and errors for the supported web services protocols.

Service Configuration

defaultService :: Service

API version 2018-05-14 of the Amazon IoT 1-Click Devices Service SDK configuration.

Errors

Error matchers are designed for use with the functions provided by Control.Exception.Lens. This allows catching (and rethrowing) service specific errors returned by IoT1ClickDevices.

InvalidRequestException

_InvalidRequestException :: AsError a => Getting (First ServiceError) a ServiceError

Prism for InvalidRequestException' errors.

PreconditionFailedException

_PreconditionFailedException :: AsError a => Getting (First ServiceError) a ServiceError

Prism for PreconditionFailedException' errors.

ForbiddenException

_ForbiddenException :: AsError a => Getting (First ServiceError) a ServiceError

Prism for ForbiddenException' errors.

RangeNotSatisfiableException

_RangeNotSatisfiableException :: AsError a => Getting (First ServiceError) a ServiceError

Prism for RangeNotSatisfiableException' errors.

InternalFailureException

_InternalFailureException :: AsError a => Getting (First ServiceError) a ServiceError

Prism for InternalFailureException' errors.

ResourceConflictException

_ResourceConflictException :: AsError a => Getting (First ServiceError) a ServiceError

Prism for ResourceConflictException' errors.

ResourceNotFoundException

_ResourceNotFoundException :: AsError a => Getting (First ServiceError) a ServiceError

Prism for ResourceNotFoundException' errors.

Waiters

Waiters poll by repeatedly sending a request until some remote success condition configured by the Wait specification is fulfilled. The Wait specification determines how many attempts should be made, in addition to delay and retry strategies.

Operations

Some AWS operations return results that are incomplete and require subsequent requests in order to obtain the entire result set. The process of sending subsequent requests to continue where a previous request left off is called pagination. For example, the ListObjects operation of Amazon S3 returns up to 1000 objects at a time, and you must send subsequent requests with the appropriate Marker in order to retrieve the next page of results.

Operations that have an AWSPager instance can transparently perform subsequent requests, correctly setting Markers and other request facets to iterate through the entire result set of a truncated API operation. Operations which support this have an additional note in the documentation.

Many operations have the ability to filter results on the server side. See the individual operation parameters for details.

GetDeviceMethods

data GetDeviceMethods

See: newGetDeviceMethods smart constructor.

Constructors

GetDeviceMethods' Text

newGetDeviceMethods

Arguments

:: Text
$sel:deviceId:GetDeviceMethods'

-> GetDeviceMethods

Create a value of GetDeviceMethods with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:deviceId:GetDeviceMethods', getDeviceMethods_deviceId - The unique identifier of the device.

data GetDeviceMethodsResponse

See: newGetDeviceMethodsResponse smart constructor.

Constructors

GetDeviceMethodsResponse' (Maybe [DeviceMethod]) Int

newGetDeviceMethodsResponse

Arguments

:: Int
$sel:httpStatus:GetDeviceMethodsResponse'

-> GetDeviceMethodsResponse

Create a value of GetDeviceMethodsResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:deviceMethods:GetDeviceMethodsResponse', getDeviceMethodsResponse_deviceMethods - List of available device APIs.

$sel:httpStatus:GetDeviceMethodsResponse', getDeviceMethodsResponse_httpStatus - The response's http status code.

ListTagsForResource

data ListTagsForResource

See: newListTagsForResource smart constructor.

Constructors

ListTagsForResource' Text

newListTagsForResource

Arguments

:: Text
$sel:resourceArn:ListTagsForResource'

-> ListTagsForResource

Create a value of ListTagsForResource with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:resourceArn:ListTagsForResource', listTagsForResource_resourceArn - The ARN of the resource.

data ListTagsForResourceResponse

See: newListTagsForResourceResponse smart constructor.

Constructors

ListTagsForResourceResponse' (Maybe (HashMap Text Text)) Int

newListTagsForResourceResponse

Arguments

:: Int
$sel:httpStatus:ListTagsForResourceResponse'

-> ListTagsForResourceResponse

Create a value of ListTagsForResourceResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:tags:ListTagsForResourceResponse', listTagsForResourceResponse_tags - A collection of key/value pairs defining the resource tags. For example, { "tags": {"key1": "value1", "key2": "value2"} }. For more information, see AWS Tagging Strategies.

$sel:httpStatus:ListTagsForResourceResponse', listTagsForResourceResponse_httpStatus - The response's http status code.

ClaimDevicesByClaimCode

data ClaimDevicesByClaimCode

See: newClaimDevicesByClaimCode smart constructor.

Constructors

ClaimDevicesByClaimCode' Text

newClaimDevicesByClaimCode

Arguments

:: Text
$sel:claimCode:ClaimDevicesByClaimCode'

-> ClaimDevicesByClaimCode

Create a value of ClaimDevicesByClaimCode with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:claimCode:ClaimDevicesByClaimCode', claimDevicesByClaimCode_claimCode - The claim code, starting with "C-", as provided by the device manufacturer.

data ClaimDevicesByClaimCodeResponse

See: newClaimDevicesByClaimCodeResponse smart constructor.

Constructors

ClaimDevicesByClaimCodeResponse' (Maybe Text) (Maybe Int) Int

newClaimDevicesByClaimCodeResponse

Arguments

:: Int
$sel:httpStatus:ClaimDevicesByClaimCodeResponse'

-> ClaimDevicesByClaimCodeResponse

Create a value of ClaimDevicesByClaimCodeResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:claimCode:ClaimDevicesByClaimCode', claimDevicesByClaimCodeResponse_claimCode - The claim code provided by the device manufacturer.

$sel:total:ClaimDevicesByClaimCodeResponse', claimDevicesByClaimCodeResponse_total - The total number of devices associated with the claim code that has been processed in the claim request.

$sel:httpStatus:ClaimDevicesByClaimCodeResponse', claimDevicesByClaimCodeResponse_httpStatus - The response's http status code.

InitiateDeviceClaim

data InitiateDeviceClaim

See: newInitiateDeviceClaim smart constructor.

Constructors

InitiateDeviceClaim' Text

newInitiateDeviceClaim

Arguments

:: Text
$sel:deviceId:InitiateDeviceClaim'

-> InitiateDeviceClaim

Create a value of InitiateDeviceClaim with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:deviceId:InitiateDeviceClaim', initiateDeviceClaim_deviceId - The unique identifier of the device.

data InitiateDeviceClaimResponse

See: newInitiateDeviceClaimResponse smart constructor.

Constructors

InitiateDeviceClaimResponse' (Maybe Text) Int

newInitiateDeviceClaimResponse

Arguments

:: Int
$sel:httpStatus:InitiateDeviceClaimResponse'

-> InitiateDeviceClaimResponse

Create a value of InitiateDeviceClaimResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:state:InitiateDeviceClaimResponse', initiateDeviceClaimResponse_state - The device's final claim state.

$sel:httpStatus:InitiateDeviceClaimResponse', initiateDeviceClaimResponse_httpStatus - The response's http status code.

InvokeDeviceMethod

data InvokeDeviceMethod

See: newInvokeDeviceMethod smart constructor.

Constructors

InvokeDeviceMethod' (Maybe Text) (Maybe DeviceMethod) Text

newInvokeDeviceMethod

Arguments

:: Text
$sel:deviceId:InvokeDeviceMethod'

-> InvokeDeviceMethod

Create a value of InvokeDeviceMethod with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:deviceMethodParameters:InvokeDeviceMethod', invokeDeviceMethod_deviceMethodParameters - A JSON encoded string containing the device method request parameters.

$sel:deviceMethod:InvokeDeviceMethod', invokeDeviceMethod_deviceMethod - The device method to invoke.

$sel:deviceId:InvokeDeviceMethod', invokeDeviceMethod_deviceId - The unique identifier of the device.

data InvokeDeviceMethodResponse

See: newInvokeDeviceMethodResponse smart constructor.

Constructors

InvokeDeviceMethodResponse' (Maybe Text) Int

newInvokeDeviceMethodResponse

Arguments

:: Int
$sel:httpStatus:InvokeDeviceMethodResponse'

-> InvokeDeviceMethodResponse

Create a value of InvokeDeviceMethodResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:deviceMethodResponse:InvokeDeviceMethodResponse', invokeDeviceMethodResponse_deviceMethodResponse - A JSON encoded string containing the device method response.

$sel:httpStatus:InvokeDeviceMethodResponse', invokeDeviceMethodResponse_httpStatus - The response's http status code.

DescribeDevice

data DescribeDevice

See: newDescribeDevice smart constructor.

Constructors

DescribeDevice' Text

newDescribeDevice

Arguments

:: Text
$sel:deviceId:DescribeDevice'

-> DescribeDevice

Create a value of DescribeDevice with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:deviceId:DescribeDevice', describeDevice_deviceId - The unique identifier of the device.

data DescribeDeviceResponse

See: newDescribeDeviceResponse smart constructor.

Constructors

DescribeDeviceResponse' (Maybe DeviceDescription) Int

newDescribeDeviceResponse

Arguments

:: Int
$sel:httpStatus:DescribeDeviceResponse'

-> DescribeDeviceResponse

Create a value of DescribeDeviceResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:deviceDescription:DescribeDeviceResponse', describeDeviceResponse_deviceDescription - Device details.

$sel:httpStatus:DescribeDeviceResponse', describeDeviceResponse_httpStatus - The response's http status code.

ListDeviceEvents (Paginated)

data ListDeviceEvents

See: newListDeviceEvents smart constructor.

Constructors

ListDeviceEvents' (Maybe Text) (Maybe Natural) Text POSIX POSIX

newListDeviceEvents

Arguments

:: Text
$sel:deviceId:ListDeviceEvents'

-> UTCTime
$sel:fromTimeStamp:ListDeviceEvents'

-> UTCTime
$sel:toTimeStamp:ListDeviceEvents'

-> ListDeviceEvents

Create a value of ListDeviceEvents with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:nextToken:ListDeviceEvents', listDeviceEvents_nextToken - The token to retrieve the next set of results.

$sel:maxResults:ListDeviceEvents', listDeviceEvents_maxResults - The maximum number of results to return per request. If not set, a default value of 100 is used.

$sel:deviceId:ListDeviceEvents', listDeviceEvents_deviceId - The unique identifier of the device.

$sel:fromTimeStamp:ListDeviceEvents', listDeviceEvents_fromTimeStamp - The start date for the device event query, in ISO8061 format. For example, 2018-03-28T15:45:12.880Z

$sel:toTimeStamp:ListDeviceEvents', listDeviceEvents_toTimeStamp - The end date for the device event query, in ISO8061 format. For example, 2018-03-28T15:45:12.880Z

data ListDeviceEventsResponse

See: newListDeviceEventsResponse smart constructor.

Constructors

ListDeviceEventsResponse' (Maybe Text) (Maybe [DeviceEvent]) Int

newListDeviceEventsResponse

Arguments

:: Int
$sel:httpStatus:ListDeviceEventsResponse'

-> ListDeviceEventsResponse

Create a value of ListDeviceEventsResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:nextToken:ListDeviceEvents', listDeviceEventsResponse_nextToken - The token to retrieve the next set of results.

$sel:events:ListDeviceEventsResponse', listDeviceEventsResponse_events - An array of zero or more elements describing the event(s) associated with the device.

$sel:httpStatus:ListDeviceEventsResponse', listDeviceEventsResponse_httpStatus - The response's http status code.

FinalizeDeviceClaim

data FinalizeDeviceClaim

See: newFinalizeDeviceClaim smart constructor.

Constructors

FinalizeDeviceClaim' (Maybe (HashMap Text Text)) Text

newFinalizeDeviceClaim

Arguments

:: Text
$sel:deviceId:FinalizeDeviceClaim'

-> FinalizeDeviceClaim

Create a value of FinalizeDeviceClaim with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:tags:FinalizeDeviceClaim', finalizeDeviceClaim_tags - A collection of key/value pairs defining the resource tags. For example, { "tags": {"key1": "value1", "key2": "value2"} }. For more information, see AWS Tagging Strategies.

$sel:deviceId:FinalizeDeviceClaim', finalizeDeviceClaim_deviceId - The unique identifier of the device.

data FinalizeDeviceClaimResponse

See: newFinalizeDeviceClaimResponse smart constructor.

Constructors

FinalizeDeviceClaimResponse' (Maybe Text) Int

newFinalizeDeviceClaimResponse

Arguments

:: Int
$sel:httpStatus:FinalizeDeviceClaimResponse'

-> FinalizeDeviceClaimResponse

Create a value of FinalizeDeviceClaimResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:state:FinalizeDeviceClaimResponse', finalizeDeviceClaimResponse_state - The device's final claim state.

$sel:httpStatus:FinalizeDeviceClaimResponse', finalizeDeviceClaimResponse_httpStatus - The response's http status code.

TagResource

data TagResource

See: newTagResource smart constructor.

Constructors

TagResource' Text (HashMap Text Text)

newTagResource

Arguments

:: Text
$sel:resourceArn:TagResource'

-> TagResource

Create a value of TagResource with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:resourceArn:TagResource', tagResource_resourceArn - The ARN of the resource.

$sel:tags:TagResource', tagResource_tags - A collection of key/value pairs defining the resource tags. For example, { "tags": {"key1": "value1", "key2": "value2"} }. For more information, see AWS Tagging Strategies.

data TagResourceResponse

See: newTagResourceResponse smart constructor.

Constructors

TagResourceResponse'

newTagResourceResponse :: TagResourceResponse

Create a value of TagResourceResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

UntagResource

data UntagResource

See: newUntagResource smart constructor.

Constructors

UntagResource' [Text] Text

newUntagResource

Arguments

:: Text
$sel:resourceArn:UntagResource'

-> UntagResource

Create a value of UntagResource with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:tagKeys:UntagResource', untagResource_tagKeys - A collections of tag keys. For example, {"key1","key2"}

$sel:resourceArn:UntagResource', untagResource_resourceArn - The ARN of the resource.

data UntagResourceResponse

See: newUntagResourceResponse smart constructor.

Constructors

UntagResourceResponse'

newUntagResourceResponse :: UntagResourceResponse

Create a value of UntagResourceResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

UpdateDeviceState

data UpdateDeviceState

See: newUpdateDeviceState smart constructor.

Constructors

UpdateDeviceState' (Maybe Bool) Text

newUpdateDeviceState

Arguments

:: Text
$sel:deviceId:UpdateDeviceState'

-> UpdateDeviceState

Create a value of UpdateDeviceState with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:enabled:UpdateDeviceState', updateDeviceState_enabled - If true, the device is enabled. If false, the device is disabled.

$sel:deviceId:UpdateDeviceState', updateDeviceState_deviceId - The unique identifier of the device.

data UpdateDeviceStateResponse

See: newUpdateDeviceStateResponse smart constructor.

Constructors

UpdateDeviceStateResponse' Int

newUpdateDeviceStateResponse

Arguments

:: Int
$sel:httpStatus:UpdateDeviceStateResponse'

-> UpdateDeviceStateResponse

Create a value of UpdateDeviceStateResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:httpStatus:UpdateDeviceStateResponse', updateDeviceStateResponse_httpStatus - The response's http status code.

UnclaimDevice

data UnclaimDevice

See: newUnclaimDevice smart constructor.

Constructors

UnclaimDevice' Text

newUnclaimDevice

Arguments

:: Text
$sel:deviceId:UnclaimDevice'

-> UnclaimDevice

Create a value of UnclaimDevice with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:deviceId:UnclaimDevice', unclaimDevice_deviceId - The unique identifier of the device.

data UnclaimDeviceResponse

See: newUnclaimDeviceResponse smart constructor.

Constructors

UnclaimDeviceResponse' (Maybe Text) Int

newUnclaimDeviceResponse

Arguments

:: Int
$sel:httpStatus:UnclaimDeviceResponse'

-> UnclaimDeviceResponse

Create a value of UnclaimDeviceResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:state:UnclaimDeviceResponse', unclaimDeviceResponse_state - The device's final claim state.

$sel:httpStatus:UnclaimDeviceResponse', unclaimDeviceResponse_httpStatus - The response's http status code.

ListDevices (Paginated)

data ListDevices

See: newListDevices smart constructor.

Constructors

ListDevices' (Maybe Text) (Maybe Text) (Maybe Natural)

newListDevices :: ListDevices

Create a value of ListDevices with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:nextToken:ListDevices', listDevices_nextToken - The token to retrieve the next set of results.

$sel:deviceType:ListDevices', listDevices_deviceType - The type of the device, such as "button".

$sel:maxResults:ListDevices', listDevices_maxResults - The maximum number of results to return per request. If not set, a default value of 100 is used.

data ListDevicesResponse

See: newListDevicesResponse smart constructor.

Constructors

ListDevicesResponse' (Maybe Text) (Maybe [DeviceDescription]) Int

newListDevicesResponse

Arguments

:: Int
$sel:httpStatus:ListDevicesResponse'

-> ListDevicesResponse

Create a value of ListDevicesResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:nextToken:ListDevices', listDevicesResponse_nextToken - The token to retrieve the next set of results.

$sel:devices:ListDevicesResponse', listDevicesResponse_devices - A list of devices.

$sel:httpStatus:ListDevicesResponse', listDevicesResponse_httpStatus - The response's http status code.

Types

Attributes

data Attributes

See: newAttributes smart constructor.

Constructors

Attributes'

newAttributes :: Attributes

Create a value of Attributes with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

Device

data Device

See: newDevice smart constructor.

Constructors

Device' (Maybe Attributes) (Maybe Text) (Maybe Text)

newDevice :: Device

Create a value of Device with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:attributes:Device', device_attributes - The user specified attributes associated with the device for an event.

$sel:deviceId:Device', device_deviceId - The unique identifier of the device.

$sel:type':Device', device_type - The device type, such as "button".

DeviceDescription

data DeviceDescription

See: newDeviceDescription smart constructor.

Constructors

DeviceDescription' (Maybe Double) (Maybe Bool) (Maybe Text) (Maybe (HashMap Text Text)) (Maybe Text) (Maybe Text) (Maybe (HashMap Text Text))

newDeviceDescription :: DeviceDescription

Create a value of DeviceDescription with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:remainingLife:DeviceDescription', deviceDescription_remainingLife - A value between 0 and 1 inclusive, representing the fraction of life remaining for the device.

$sel:enabled:DeviceDescription', deviceDescription_enabled - A Boolean value indicating whether or not the device is enabled.

$sel:arn:DeviceDescription', deviceDescription_arn - The ARN of the device.

$sel:attributes:DeviceDescription', deviceDescription_attributes - An array of zero or more elements of DeviceAttribute objects providing user specified device attributes.

$sel:deviceId:DeviceDescription', deviceDescription_deviceId - The unique identifier of the device.

$sel:type':DeviceDescription', deviceDescription_type - The type of the device, such as "button".

$sel:tags:DeviceDescription', deviceDescription_tags - The tags currently associated with the AWS IoT 1-Click device.

DeviceEvent

data DeviceEvent

See: newDeviceEvent smart constructor.

Constructors

DeviceEvent' (Maybe Text) (Maybe Device)

newDeviceEvent :: DeviceEvent

Create a value of DeviceEvent with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:stdEvent:DeviceEvent', deviceEvent_stdEvent - A serialized JSON object representing the device-type specific event.

$sel:device:DeviceEvent', deviceEvent_device - An object representing the device associated with the event.

DeviceMethod

data DeviceMethod

See: newDeviceMethod smart constructor.

Constructors

DeviceMethod' (Maybe Text) (Maybe Text)

newDeviceMethod :: DeviceMethod

Create a value of DeviceMethod with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:methodName:DeviceMethod', deviceMethod_methodName - The name of the method applicable to the deviceType.

$sel:deviceType:DeviceMethod', deviceMethod_deviceType - The type of the device, such as "button".