Amazonka.IoT1ClickDevices.DescribeDevice

Description

Given a device ID, returns a DescribeDeviceResponse object describing the details of the device.

Creating a Request

data DescribeDevice

See: newDescribeDevice smart constructor.

Constructors

DescribeDevice'

Fields

deviceId :: Text

The unique identifier of the device.

newDescribeDevice

Arguments

:: Text
$sel:deviceId:DescribeDevice'

-> DescribeDevice

Create a value of DescribeDevice with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:deviceId:DescribeDevice', describeDevice_deviceId - The unique identifier of the device.

Request Lenses

describeDevice_deviceId :: Lens' DescribeDevice Text

The unique identifier of the device.

Destructuring the Response

data DescribeDeviceResponse

See: newDescribeDeviceResponse smart constructor.

Constructors

DescribeDeviceResponse'

Fields

deviceDescription :: Maybe DeviceDescription

Device details.

httpStatus :: Int

The response's http status code.

newDescribeDeviceResponse

Arguments

:: Int
$sel:httpStatus:DescribeDeviceResponse'

-> DescribeDeviceResponse

Create a value of DescribeDeviceResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:deviceDescription:DescribeDeviceResponse', describeDeviceResponse_deviceDescription - Device details.

$sel:httpStatus:DescribeDeviceResponse', describeDeviceResponse_httpStatus - The response's http status code.

Response Lenses

describeDeviceResponse_deviceDescription :: Lens' DescribeDeviceResponse (Maybe DeviceDescription)

Device details.

describeDeviceResponse_httpStatus :: Lens' DescribeDeviceResponse Int

The response's http status code.