Amazonka.IoT1ClickDevices.GetDeviceMethods

Description

Given a device ID, returns the invokable methods associated with the device.

Creating a Request

data GetDeviceMethods

See: newGetDeviceMethods smart constructor.

Constructors

GetDeviceMethods'

Fields

deviceId :: Text

The unique identifier of the device.

newGetDeviceMethods

Arguments

:: Text
$sel:deviceId:GetDeviceMethods'

-> GetDeviceMethods

Create a value of GetDeviceMethods with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:deviceId:GetDeviceMethods', getDeviceMethods_deviceId - The unique identifier of the device.

Request Lenses

getDeviceMethods_deviceId :: Lens' GetDeviceMethods Text

The unique identifier of the device.

Destructuring the Response

data GetDeviceMethodsResponse

See: newGetDeviceMethodsResponse smart constructor.

Constructors

GetDeviceMethodsResponse'

Fields

deviceMethods :: Maybe [DeviceMethod]

List of available device APIs.

httpStatus :: Int

The response's http status code.

newGetDeviceMethodsResponse

Arguments

:: Int
$sel:httpStatus:GetDeviceMethodsResponse'

-> GetDeviceMethodsResponse

Create a value of GetDeviceMethodsResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:deviceMethods:GetDeviceMethodsResponse', getDeviceMethodsResponse_deviceMethods - List of available device APIs.

$sel:httpStatus:GetDeviceMethodsResponse', getDeviceMethodsResponse_httpStatus - The response's http status code.

Response Lenses

getDeviceMethodsResponse_deviceMethods :: Lens' GetDeviceMethodsResponse (Maybe [DeviceMethod])

List of available device APIs.

getDeviceMethodsResponse_httpStatus :: Lens' GetDeviceMethodsResponse Int

The response's http status code.