Amazonka.IoT1ClickDevices.Lens

Description

Operations

GetDeviceMethods

getDeviceMethods_deviceId :: Lens' GetDeviceMethods Text

The unique identifier of the device.

getDeviceMethodsResponse_deviceMethods :: Lens' GetDeviceMethodsResponse (Maybe [DeviceMethod])

List of available device APIs.

getDeviceMethodsResponse_httpStatus :: Lens' GetDeviceMethodsResponse Int

The response's http status code.

ListTagsForResource

listTagsForResource_resourceArn :: Lens' ListTagsForResource Text

The ARN of the resource.

listTagsForResourceResponse_tags :: Lens' ListTagsForResourceResponse (Maybe (HashMap Text Text))

A collection of key/value pairs defining the resource tags. For example, { "tags": {"key1": "value1", "key2": "value2"} }. For more information, see AWS Tagging Strategies.

listTagsForResourceResponse_httpStatus :: Lens' ListTagsForResourceResponse Int

The response's http status code.

ClaimDevicesByClaimCode

claimDevicesByClaimCode_claimCode :: Lens' ClaimDevicesByClaimCode Text

The claim code, starting with "C-", as provided by the device manufacturer.

claimDevicesByClaimCodeResponse_claimCode :: Lens' ClaimDevicesByClaimCodeResponse (Maybe Text)

The claim code provided by the device manufacturer.

claimDevicesByClaimCodeResponse_total :: Lens' ClaimDevicesByClaimCodeResponse (Maybe Int)

The total number of devices associated with the claim code that has been processed in the claim request.

claimDevicesByClaimCodeResponse_httpStatus :: Lens' ClaimDevicesByClaimCodeResponse Int

The response's http status code.

InitiateDeviceClaim

initiateDeviceClaim_deviceId :: Lens' InitiateDeviceClaim Text

The unique identifier of the device.

initiateDeviceClaimResponse_state :: Lens' InitiateDeviceClaimResponse (Maybe Text)

The device's final claim state.

initiateDeviceClaimResponse_httpStatus :: Lens' InitiateDeviceClaimResponse Int

The response's http status code.

InvokeDeviceMethod

invokeDeviceMethod_deviceMethodParameters :: Lens' InvokeDeviceMethod (Maybe Text)

A JSON encoded string containing the device method request parameters.

invokeDeviceMethod_deviceMethod :: Lens' InvokeDeviceMethod (Maybe DeviceMethod)

The device method to invoke.

invokeDeviceMethod_deviceId :: Lens' InvokeDeviceMethod Text

The unique identifier of the device.

invokeDeviceMethodResponse_deviceMethodResponse :: Lens' InvokeDeviceMethodResponse (Maybe Text)

A JSON encoded string containing the device method response.

invokeDeviceMethodResponse_httpStatus :: Lens' InvokeDeviceMethodResponse Int

The response's http status code.

DescribeDevice

describeDevice_deviceId :: Lens' DescribeDevice Text

The unique identifier of the device.

describeDeviceResponse_deviceDescription :: Lens' DescribeDeviceResponse (Maybe DeviceDescription)

Device details.

describeDeviceResponse_httpStatus :: Lens' DescribeDeviceResponse Int

The response's http status code.

ListDeviceEvents

listDeviceEvents_nextToken :: Lens' ListDeviceEvents (Maybe Text)

The token to retrieve the next set of results.

listDeviceEvents_maxResults :: Lens' ListDeviceEvents (Maybe Natural)

The maximum number of results to return per request. If not set, a default value of 100 is used.

listDeviceEvents_deviceId :: Lens' ListDeviceEvents Text

The unique identifier of the device.

listDeviceEvents_fromTimeStamp :: Lens' ListDeviceEvents UTCTime

The start date for the device event query, in ISO8061 format. For example, 2018-03-28T15:45:12.880Z

listDeviceEvents_toTimeStamp :: Lens' ListDeviceEvents UTCTime

The end date for the device event query, in ISO8061 format. For example, 2018-03-28T15:45:12.880Z

listDeviceEventsResponse_nextToken :: Lens' ListDeviceEventsResponse (Maybe Text)

The token to retrieve the next set of results.

listDeviceEventsResponse_events :: Lens' ListDeviceEventsResponse (Maybe [DeviceEvent])

An array of zero or more elements describing the event(s) associated with the device.

listDeviceEventsResponse_httpStatus :: Lens' ListDeviceEventsResponse Int

The response's http status code.

FinalizeDeviceClaim

finalizeDeviceClaim_tags :: Lens' FinalizeDeviceClaim (Maybe (HashMap Text Text))

A collection of key/value pairs defining the resource tags. For example, { "tags": {"key1": "value1", "key2": "value2"} }. For more information, see AWS Tagging Strategies.

finalizeDeviceClaim_deviceId :: Lens' FinalizeDeviceClaim Text

The unique identifier of the device.

finalizeDeviceClaimResponse_state :: Lens' FinalizeDeviceClaimResponse (Maybe Text)

The device's final claim state.

finalizeDeviceClaimResponse_httpStatus :: Lens' FinalizeDeviceClaimResponse Int

The response's http status code.

TagResource

tagResource_resourceArn :: Lens' TagResource Text

The ARN of the resource.

tagResource_tags :: Lens' TagResource (HashMap Text Text)

A collection of key/value pairs defining the resource tags. For example, { "tags": {"key1": "value1", "key2": "value2"} }. For more information, see AWS Tagging Strategies.

UntagResource

untagResource_tagKeys :: Lens' UntagResource [Text]

A collections of tag keys. For example, {"key1","key2"}

untagResource_resourceArn :: Lens' UntagResource Text

The ARN of the resource.

UpdateDeviceState

updateDeviceState_enabled :: Lens' UpdateDeviceState (Maybe Bool)

If true, the device is enabled. If false, the device is disabled.

updateDeviceState_deviceId :: Lens' UpdateDeviceState Text

The unique identifier of the device.

updateDeviceStateResponse_httpStatus :: Lens' UpdateDeviceStateResponse Int

The response's http status code.

UnclaimDevice

unclaimDevice_deviceId :: Lens' UnclaimDevice Text

The unique identifier of the device.

unclaimDeviceResponse_state :: Lens' UnclaimDeviceResponse (Maybe Text)

The device's final claim state.

unclaimDeviceResponse_httpStatus :: Lens' UnclaimDeviceResponse Int

The response's http status code.

ListDevices

listDevices_nextToken :: Lens' ListDevices (Maybe Text)

The token to retrieve the next set of results.

listDevices_deviceType :: Lens' ListDevices (Maybe Text)

The type of the device, such as "button".

listDevices_maxResults :: Lens' ListDevices (Maybe Natural)

The maximum number of results to return per request. If not set, a default value of 100 is used.

listDevicesResponse_nextToken :: Lens' ListDevicesResponse (Maybe Text)

The token to retrieve the next set of results.

listDevicesResponse_devices :: Lens' ListDevicesResponse (Maybe [DeviceDescription])

A list of devices.

listDevicesResponse_httpStatus :: Lens' ListDevicesResponse Int

The response's http status code.

Types

Attributes

Device

device_attributes :: Lens' Device (Maybe Attributes)

The user specified attributes associated with the device for an event.

device_deviceId :: Lens' Device (Maybe Text)

The unique identifier of the device.

device_type :: Lens' Device (Maybe Text)

The device type, such as "button".

DeviceDescription

deviceDescription_remainingLife :: Lens' DeviceDescription (Maybe Double)

A value between 0 and 1 inclusive, representing the fraction of life remaining for the device.

deviceDescription_enabled :: Lens' DeviceDescription (Maybe Bool)

A Boolean value indicating whether or not the device is enabled.

deviceDescription_arn :: Lens' DeviceDescription (Maybe Text)

The ARN of the device.

deviceDescription_attributes :: Lens' DeviceDescription (Maybe (HashMap Text Text))

An array of zero or more elements of DeviceAttribute objects providing user specified device attributes.

deviceDescription_deviceId :: Lens' DeviceDescription (Maybe Text)

The unique identifier of the device.

deviceDescription_type :: Lens' DeviceDescription (Maybe Text)

The type of the device, such as "button".

deviceDescription_tags :: Lens' DeviceDescription (Maybe (HashMap Text Text))

The tags currently associated with the AWS IoT 1-Click device.

DeviceEvent

deviceEvent_stdEvent :: Lens' DeviceEvent (Maybe Text)

A serialized JSON object representing the device-type specific event.

deviceEvent_device :: Lens' DeviceEvent (Maybe Device)

An object representing the device associated with the event.

DeviceMethod

deviceMethod_methodName :: Lens' DeviceMethod (Maybe Text)

The name of the method applicable to the deviceType.

deviceMethod_deviceType :: Lens' DeviceMethod (Maybe Text)

The type of the device, such as "button".