Amazonka.IoT1ClickDevices.Types

Description

Service Configuration

defaultService :: Service

API version 2018-05-14 of the Amazon IoT 1-Click Devices Service SDK configuration.

Errors

_InvalidRequestException :: AsError a => Getting (First ServiceError) a ServiceError

Prism for InvalidRequestException' errors.

_PreconditionFailedException :: AsError a => Getting (First ServiceError) a ServiceError

Prism for PreconditionFailedException' errors.

_ForbiddenException :: AsError a => Getting (First ServiceError) a ServiceError

Prism for ForbiddenException' errors.

_RangeNotSatisfiableException :: AsError a => Getting (First ServiceError) a ServiceError

Prism for RangeNotSatisfiableException' errors.

_InternalFailureException :: AsError a => Getting (First ServiceError) a ServiceError

Prism for InternalFailureException' errors.

_ResourceConflictException :: AsError a => Getting (First ServiceError) a ServiceError

Prism for ResourceConflictException' errors.

_ResourceNotFoundException :: AsError a => Getting (First ServiceError) a ServiceError

Prism for ResourceNotFoundException' errors.

Attributes

data Attributes

See: newAttributes smart constructor.

Constructors

Attributes'

newAttributes :: Attributes

Create a value of Attributes with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

Device

data Device

See: newDevice smart constructor.

Constructors

Device'

Fields

attributes :: Maybe Attributes

The user specified attributes associated with the device for an event.

deviceId :: Maybe Text

The unique identifier of the device.

type' :: Maybe Text

The device type, such as "button".

newDevice :: Device

Create a value of Device with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:attributes:Device', device_attributes - The user specified attributes associated with the device for an event.

$sel:deviceId:Device', device_deviceId - The unique identifier of the device.

$sel:type':Device', device_type - The device type, such as "button".

device_attributes :: Lens' Device (Maybe Attributes)

The user specified attributes associated with the device for an event.

device_deviceId :: Lens' Device (Maybe Text)

The unique identifier of the device.

device_type :: Lens' Device (Maybe Text)

The device type, such as "button".

DeviceDescription

data DeviceDescription

See: newDeviceDescription smart constructor.

Constructors

DeviceDescription'

Fields

remainingLife :: Maybe Double

A value between 0 and 1 inclusive, representing the fraction of life remaining for the device.

enabled :: Maybe Bool

A Boolean value indicating whether or not the device is enabled.

arn :: Maybe Text

The ARN of the device.

attributes :: Maybe (HashMap Text Text)

An array of zero or more elements of DeviceAttribute objects providing user specified device attributes.

deviceId :: Maybe Text

The unique identifier of the device.

type' :: Maybe Text

The type of the device, such as "button".

tags :: Maybe (HashMap Text Text)

The tags currently associated with the AWS IoT 1-Click device.

newDeviceDescription :: DeviceDescription

Create a value of DeviceDescription with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:remainingLife:DeviceDescription', deviceDescription_remainingLife - A value between 0 and 1 inclusive, representing the fraction of life remaining for the device.

$sel:enabled:DeviceDescription', deviceDescription_enabled - A Boolean value indicating whether or not the device is enabled.

$sel:arn:DeviceDescription', deviceDescription_arn - The ARN of the device.

$sel:attributes:DeviceDescription', deviceDescription_attributes - An array of zero or more elements of DeviceAttribute objects providing user specified device attributes.

$sel:deviceId:DeviceDescription', deviceDescription_deviceId - The unique identifier of the device.

$sel:type':DeviceDescription', deviceDescription_type - The type of the device, such as "button".

$sel:tags:DeviceDescription', deviceDescription_tags - The tags currently associated with the AWS IoT 1-Click device.

deviceDescription_remainingLife :: Lens' DeviceDescription (Maybe Double)

A value between 0 and 1 inclusive, representing the fraction of life remaining for the device.

deviceDescription_enabled :: Lens' DeviceDescription (Maybe Bool)

A Boolean value indicating whether or not the device is enabled.

deviceDescription_arn :: Lens' DeviceDescription (Maybe Text)

The ARN of the device.

deviceDescription_attributes :: Lens' DeviceDescription (Maybe (HashMap Text Text))

An array of zero or more elements of DeviceAttribute objects providing user specified device attributes.

deviceDescription_deviceId :: Lens' DeviceDescription (Maybe Text)

The unique identifier of the device.

deviceDescription_type :: Lens' DeviceDescription (Maybe Text)

The type of the device, such as "button".

deviceDescription_tags :: Lens' DeviceDescription (Maybe (HashMap Text Text))

The tags currently associated with the AWS IoT 1-Click device.

DeviceEvent

data DeviceEvent

See: newDeviceEvent smart constructor.

Constructors

DeviceEvent'

Fields

stdEvent :: Maybe Text

A serialized JSON object representing the device-type specific event.

device :: Maybe Device

An object representing the device associated with the event.

newDeviceEvent :: DeviceEvent

Create a value of DeviceEvent with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:stdEvent:DeviceEvent', deviceEvent_stdEvent - A serialized JSON object representing the device-type specific event.

$sel:device:DeviceEvent', deviceEvent_device - An object representing the device associated with the event.

deviceEvent_stdEvent :: Lens' DeviceEvent (Maybe Text)

A serialized JSON object representing the device-type specific event.

deviceEvent_device :: Lens' DeviceEvent (Maybe Device)

An object representing the device associated with the event.

DeviceMethod

data DeviceMethod

See: newDeviceMethod smart constructor.

Constructors

DeviceMethod'

Fields

methodName :: Maybe Text

The name of the method applicable to the deviceType.

deviceType :: Maybe Text

The type of the device, such as "button".

newDeviceMethod :: DeviceMethod

Create a value of DeviceMethod with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:methodName:DeviceMethod', deviceMethod_methodName - The name of the method applicable to the deviceType.

$sel:deviceType:DeviceMethod', deviceMethod_deviceType - The type of the device, such as "button".

deviceMethod_methodName :: Lens' DeviceMethod (Maybe Text)

The name of the method applicable to the deviceType.

deviceMethod_deviceType :: Lens' DeviceMethod (Maybe Text)

The type of the device, such as "button".