Amazonka.IoT1ClickDevices.Types.DeviceDescription

Description

Documentation

data DeviceDescription

See: newDeviceDescription smart constructor.

Constructors

DeviceDescription'

Fields

remainingLife :: Maybe Double

A value between 0 and 1 inclusive, representing the fraction of life remaining for the device.

enabled :: Maybe Bool

A Boolean value indicating whether or not the device is enabled.

arn :: Maybe Text

The ARN of the device.

attributes :: Maybe (HashMap Text Text)

An array of zero or more elements of DeviceAttribute objects providing user specified device attributes.

deviceId :: Maybe Text

The unique identifier of the device.

type' :: Maybe Text

The type of the device, such as "button".

tags :: Maybe (HashMap Text Text)

The tags currently associated with the AWS IoT 1-Click device.

newDeviceDescription :: DeviceDescription

Create a value of DeviceDescription with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:remainingLife:DeviceDescription', deviceDescription_remainingLife - A value between 0 and 1 inclusive, representing the fraction of life remaining for the device.

$sel:enabled:DeviceDescription', deviceDescription_enabled - A Boolean value indicating whether or not the device is enabled.

$sel:arn:DeviceDescription', deviceDescription_arn - The ARN of the device.

$sel:attributes:DeviceDescription', deviceDescription_attributes - An array of zero or more elements of DeviceAttribute objects providing user specified device attributes.

$sel:deviceId:DeviceDescription', deviceDescription_deviceId - The unique identifier of the device.

$sel:type':DeviceDescription', deviceDescription_type - The type of the device, such as "button".

$sel:tags:DeviceDescription', deviceDescription_tags - The tags currently associated with the AWS IoT 1-Click device.

deviceDescription_remainingLife :: Lens' DeviceDescription (Maybe Double)

A value between 0 and 1 inclusive, representing the fraction of life remaining for the device.

deviceDescription_enabled :: Lens' DeviceDescription (Maybe Bool)

A Boolean value indicating whether or not the device is enabled.

deviceDescription_arn :: Lens' DeviceDescription (Maybe Text)

The ARN of the device.

deviceDescription_attributes :: Lens' DeviceDescription (Maybe (HashMap Text Text))

An array of zero or more elements of DeviceAttribute objects providing user specified device attributes.

deviceDescription_deviceId :: Lens' DeviceDescription (Maybe Text)

The unique identifier of the device.

deviceDescription_type :: Lens' DeviceDescription (Maybe Text)

The type of the device, such as "button".

deviceDescription_tags :: Lens' DeviceDescription (Maybe (HashMap Text Text))

The tags currently associated with the AWS IoT 1-Click device.