Amazonka.IoT1ClickDevices.UpdateDeviceState

Description

Using a Boolean value (true or false), this operation enables or disables the device given a device ID.

Creating a Request

data UpdateDeviceState

See: newUpdateDeviceState smart constructor.

Constructors

UpdateDeviceState'

Fields

enabled :: Maybe Bool

If true, the device is enabled. If false, the device is disabled.

deviceId :: Text

The unique identifier of the device.

newUpdateDeviceState

Arguments

:: Text
$sel:deviceId:UpdateDeviceState'

-> UpdateDeviceState

Create a value of UpdateDeviceState with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:enabled:UpdateDeviceState', updateDeviceState_enabled - If true, the device is enabled. If false, the device is disabled.

$sel:deviceId:UpdateDeviceState', updateDeviceState_deviceId - The unique identifier of the device.

Request Lenses

updateDeviceState_enabled :: Lens' UpdateDeviceState (Maybe Bool)

If true, the device is enabled. If false, the device is disabled.

updateDeviceState_deviceId :: Lens' UpdateDeviceState Text

The unique identifier of the device.

Destructuring the Response

data UpdateDeviceStateResponse

See: newUpdateDeviceStateResponse smart constructor.

Constructors

UpdateDeviceStateResponse'

Fields

httpStatus :: Int

The response's http status code.

newUpdateDeviceStateResponse

Arguments

:: Int
$sel:httpStatus:UpdateDeviceStateResponse'

-> UpdateDeviceStateResponse

Create a value of UpdateDeviceStateResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:httpStatus:UpdateDeviceStateResponse', updateDeviceStateResponse_httpStatus - The response's http status code.

Response Lenses

updateDeviceStateResponse_httpStatus :: Lens' UpdateDeviceStateResponse Int

The response's http status code.