Amazonka.IoT1ClickProjects

Description

Derived from API version 2018-05-14 of the AWS service descriptions, licensed under Apache 2.0.

The AWS IoT 1-Click Projects API Reference

Service Configuration

defaultService :: Service

API version 2018-05-14 of the Amazon IoT 1-Click Projects Service SDK configuration.

Errors

Error matchers are designed for use with the functions provided by Control.Exception.Lens. This allows catching (and rethrowing) service specific errors returned by IoT1ClickProjects.

InvalidRequestException

_InvalidRequestException :: AsError a => Getting (First ServiceError) a ServiceError

TooManyRequestsException

_TooManyRequestsException :: AsError a => Getting (First ServiceError) a ServiceError

InternalFailureException

_InternalFailureException :: AsError a => Getting (First ServiceError) a ServiceError

ResourceConflictException

_ResourceConflictException :: AsError a => Getting (First ServiceError) a ServiceError

ResourceNotFoundException

_ResourceNotFoundException :: AsError a => Getting (First ServiceError) a ServiceError

Waiters

Waiters poll by repeatedly sending a request until some remote success condition configured by the Wait specification is fulfilled. The Wait specification determines how many attempts should be made, in addition to delay and retry strategies.

Operations

Some AWS operations return results that are incomplete and require subsequent requests in order to obtain the entire result set. The process of sending subsequent requests to continue where a previous request left off is called pagination. For example, the ListObjects operation of Amazon S3 returns up to 1000 objects at a time, and you must send subsequent requests with the appropriate Marker in order to retrieve the next page of results.

Operations that have an AWSPager instance can transparently perform subsequent requests, correctly setting Markers and other request facets to iterate through the entire result set of a truncated API operation. Operations which support this have an additional note in the documentation.

Many operations have the ability to filter results on the server side. See the individual operation parameters for details.

ListProjects (Paginated)

data ListProjects

See: newListProjects smart constructor.

Constructors

ListProjects' (Maybe Text) (Maybe Natural)

newListProjects :: ListProjects

Create a value of ListProjects with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:nextToken:ListProjects', listProjects_nextToken - The token to retrieve the next set of results.

$sel:maxResults:ListProjects', listProjects_maxResults - The maximum number of results to return per request. If not set, a default value of 100 is used.

data ListProjectsResponse

See: newListProjectsResponse smart constructor.

Constructors

ListProjectsResponse' (Maybe Text) Int [ProjectSummary]

newListProjectsResponse

Arguments

:: Int
$sel:httpStatus:ListProjectsResponse'

-> ListProjectsResponse

Create a value of ListProjectsResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:nextToken:ListProjects', listProjectsResponse_nextToken - The token used to retrieve the next set of results - will be effectively empty if there are no further results.

$sel:httpStatus:ListProjectsResponse', listProjectsResponse_httpStatus - The response's http status code.

$sel:projects:ListProjectsResponse', listProjectsResponse_projects - An object containing the list of projects.

DeleteProject

data DeleteProject

See: newDeleteProject smart constructor.

Constructors

DeleteProject' Text

newDeleteProject

Arguments

:: Text
$sel:projectName:DeleteProject'

-> DeleteProject

Create a value of DeleteProject with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:projectName:DeleteProject', deleteProject_projectName - The name of the empty project to delete.

data DeleteProjectResponse

See: newDeleteProjectResponse smart constructor.

Constructors

DeleteProjectResponse' Int

newDeleteProjectResponse

Arguments

:: Int
$sel:httpStatus:DeleteProjectResponse'

-> DeleteProjectResponse

Create a value of DeleteProjectResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:httpStatus:DeleteProjectResponse', deleteProjectResponse_httpStatus - The response's http status code.

UpdateProject

data UpdateProject

See: newUpdateProject smart constructor.

Constructors

UpdateProject' (Maybe PlacementTemplate) (Maybe Text) Text

newUpdateProject

Arguments

:: Text
$sel:projectName:UpdateProject'

-> UpdateProject

Create a value of UpdateProject with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:placementTemplate:UpdateProject', updateProject_placementTemplate - An object defining the project update. Once a project has been created, you cannot add device template names to the project. However, for a given placementTemplate, you can update the associated callbackOverrides for the device definition using this API.

$sel:description:UpdateProject', updateProject_description - An optional user-defined description for the project.

$sel:projectName:UpdateProject', updateProject_projectName - The name of the project to be updated.

data UpdateProjectResponse

See: newUpdateProjectResponse smart constructor.

Constructors

UpdateProjectResponse' Int

newUpdateProjectResponse

Arguments

:: Int
$sel:httpStatus:UpdateProjectResponse'

-> UpdateProjectResponse

Create a value of UpdateProjectResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:httpStatus:UpdateProjectResponse', updateProjectResponse_httpStatus - The response's http status code.

ListTagsForResource

data ListTagsForResource

See: newListTagsForResource smart constructor.

Constructors

ListTagsForResource' Text

newListTagsForResource

Arguments

:: Text
$sel:resourceArn:ListTagsForResource'

-> ListTagsForResource

Create a value of ListTagsForResource with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:resourceArn:ListTagsForResource', listTagsForResource_resourceArn - The ARN of the resource whose tags you want to list.

data ListTagsForResourceResponse

See: newListTagsForResourceResponse smart constructor.

Constructors

ListTagsForResourceResponse' (Maybe (HashMap Text Text)) Int

newListTagsForResourceResponse

Arguments

:: Int
$sel:httpStatus:ListTagsForResourceResponse'

-> ListTagsForResourceResponse

Create a value of ListTagsForResourceResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:tags:ListTagsForResourceResponse', listTagsForResourceResponse_tags - The tags (metadata key/value pairs) which you have assigned to the resource.

$sel:httpStatus:ListTagsForResourceResponse', listTagsForResourceResponse_httpStatus - The response's http status code.

DisassociateDeviceFromPlacement

data DisassociateDeviceFromPlacement

See: newDisassociateDeviceFromPlacement smart constructor.

Constructors

DisassociateDeviceFromPlacement' Text Text Text

newDisassociateDeviceFromPlacement

Arguments

:: Text
$sel:projectName:DisassociateDeviceFromPlacement'

-> Text
$sel:placementName:DisassociateDeviceFromPlacement'

-> Text
$sel:deviceTemplateName:DisassociateDeviceFromPlacement'

-> DisassociateDeviceFromPlacement

Create a value of DisassociateDeviceFromPlacement with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:projectName:DisassociateDeviceFromPlacement', disassociateDeviceFromPlacement_projectName - The name of the project that contains the placement.

$sel:placementName:DisassociateDeviceFromPlacement', disassociateDeviceFromPlacement_placementName - The name of the placement that the device should be removed from.

$sel:deviceTemplateName:DisassociateDeviceFromPlacement', disassociateDeviceFromPlacement_deviceTemplateName - The device ID that should be removed from the placement.

data DisassociateDeviceFromPlacementResponse

See: newDisassociateDeviceFromPlacementResponse smart constructor.

Constructors

DisassociateDeviceFromPlacementResponse' Int

newDisassociateDeviceFromPlacementResponse

Arguments

:: Int
$sel:httpStatus:DisassociateDeviceFromPlacementResponse'

-> DisassociateDeviceFromPlacementResponse

Create a value of DisassociateDeviceFromPlacementResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:httpStatus:DisassociateDeviceFromPlacementResponse', disassociateDeviceFromPlacementResponse_httpStatus - The response's http status code.

CreatePlacement

data CreatePlacement

See: newCreatePlacement smart constructor.

Constructors

CreatePlacement' (Maybe (HashMap Text Text)) Text Text

newCreatePlacement

Arguments

:: Text
$sel:placementName:CreatePlacement'

-> Text
$sel:projectName:CreatePlacement'

-> CreatePlacement

Create a value of CreatePlacement with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:attributes:CreatePlacement', createPlacement_attributes - Optional user-defined key/value pairs providing contextual data (such as location or function) for the placement.

$sel:placementName:CreatePlacement', createPlacement_placementName - The name of the placement to be created.

$sel:projectName:CreatePlacement', createPlacement_projectName - The name of the project in which to create the placement.

data CreatePlacementResponse

See: newCreatePlacementResponse smart constructor.

Constructors

CreatePlacementResponse' Int

newCreatePlacementResponse

Arguments

:: Int
$sel:httpStatus:CreatePlacementResponse'

-> CreatePlacementResponse

Create a value of CreatePlacementResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:httpStatus:CreatePlacementResponse', createPlacementResponse_httpStatus - The response's http status code.

DescribeProject

data DescribeProject

See: newDescribeProject smart constructor.

Constructors

DescribeProject' Text

newDescribeProject

Arguments

:: Text
$sel:projectName:DescribeProject'

-> DescribeProject

Create a value of DescribeProject with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:projectName:DescribeProject', describeProject_projectName - The name of the project to be described.

data DescribeProjectResponse

See: newDescribeProjectResponse smart constructor.

Constructors

DescribeProjectResponse' Int ProjectDescription

newDescribeProjectResponse

Arguments

:: Int
$sel:httpStatus:DescribeProjectResponse'

-> ProjectDescription
$sel:project:DescribeProjectResponse'

-> DescribeProjectResponse

Create a value of DescribeProjectResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:httpStatus:DescribeProjectResponse', describeProjectResponse_httpStatus - The response's http status code.

$sel:project:DescribeProjectResponse', describeProjectResponse_project - An object describing the project.

GetDevicesInPlacement

data GetDevicesInPlacement

See: newGetDevicesInPlacement smart constructor.

Constructors

GetDevicesInPlacement' Text Text

newGetDevicesInPlacement

Arguments

:: Text
$sel:projectName:GetDevicesInPlacement'

-> Text
$sel:placementName:GetDevicesInPlacement'

-> GetDevicesInPlacement

Create a value of GetDevicesInPlacement with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:projectName:GetDevicesInPlacement', getDevicesInPlacement_projectName - The name of the project containing the placement.

$sel:placementName:GetDevicesInPlacement', getDevicesInPlacement_placementName - The name of the placement to get the devices from.

data GetDevicesInPlacementResponse

See: newGetDevicesInPlacementResponse smart constructor.

Constructors

GetDevicesInPlacementResponse' Int (HashMap Text Text)

newGetDevicesInPlacementResponse

Arguments

:: Int
$sel:httpStatus:GetDevicesInPlacementResponse'

-> GetDevicesInPlacementResponse

Create a value of GetDevicesInPlacementResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:httpStatus:GetDevicesInPlacementResponse', getDevicesInPlacementResponse_httpStatus - The response's http status code.

$sel:devices:GetDevicesInPlacementResponse', getDevicesInPlacementResponse_devices - An object containing the devices (zero or more) within the placement.

DescribePlacement

data DescribePlacement

See: newDescribePlacement smart constructor.

Constructors

DescribePlacement' Text Text

newDescribePlacement

Arguments

:: Text
$sel:placementName:DescribePlacement'

-> Text
$sel:projectName:DescribePlacement'

-> DescribePlacement

Create a value of DescribePlacement with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:placementName:DescribePlacement', describePlacement_placementName - The name of the placement within a project.

$sel:projectName:DescribePlacement', describePlacement_projectName - The project containing the placement to be described.

data DescribePlacementResponse

See: newDescribePlacementResponse smart constructor.

Constructors

DescribePlacementResponse' Int PlacementDescription

newDescribePlacementResponse

Arguments

:: Int
$sel:httpStatus:DescribePlacementResponse'

-> PlacementDescription
$sel:placement:DescribePlacementResponse'

-> DescribePlacementResponse

Create a value of DescribePlacementResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:httpStatus:DescribePlacementResponse', describePlacementResponse_httpStatus - The response's http status code.

$sel:placement:DescribePlacementResponse', describePlacementResponse_placement - An object describing the placement.

AssociateDeviceWithPlacement

data AssociateDeviceWithPlacement

See: newAssociateDeviceWithPlacement smart constructor.

Constructors

AssociateDeviceWithPlacement' Text Text Text Text

newAssociateDeviceWithPlacement

Arguments

:: Text
$sel:projectName:AssociateDeviceWithPlacement'

-> Text
$sel:placementName:AssociateDeviceWithPlacement'

-> Text
$sel:deviceId:AssociateDeviceWithPlacement'

-> Text
$sel:deviceTemplateName:AssociateDeviceWithPlacement'

-> AssociateDeviceWithPlacement

Create a value of AssociateDeviceWithPlacement with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:projectName:AssociateDeviceWithPlacement', associateDeviceWithPlacement_projectName - The name of the project containing the placement in which to associate the device.

$sel:placementName:AssociateDeviceWithPlacement', associateDeviceWithPlacement_placementName - The name of the placement in which to associate the device.

$sel:deviceId:AssociateDeviceWithPlacement', associateDeviceWithPlacement_deviceId - The ID of the physical device to be associated with the given placement in the project. Note that a mandatory 4 character prefix is required for all deviceId values.

$sel:deviceTemplateName:AssociateDeviceWithPlacement', associateDeviceWithPlacement_deviceTemplateName - The device template name to associate with the device ID.

data AssociateDeviceWithPlacementResponse

See: newAssociateDeviceWithPlacementResponse smart constructor.

Constructors

AssociateDeviceWithPlacementResponse' Int

newAssociateDeviceWithPlacementResponse

Arguments

:: Int
$sel:httpStatus:AssociateDeviceWithPlacementResponse'

-> AssociateDeviceWithPlacementResponse

Create a value of AssociateDeviceWithPlacementResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:httpStatus:AssociateDeviceWithPlacementResponse', associateDeviceWithPlacementResponse_httpStatus - The response's http status code.

DeletePlacement

data DeletePlacement

See: newDeletePlacement smart constructor.

Constructors

DeletePlacement' Text Text

newDeletePlacement

Arguments

:: Text
$sel:placementName:DeletePlacement'

-> Text
$sel:projectName:DeletePlacement'

-> DeletePlacement

Create a value of DeletePlacement with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:placementName:DeletePlacement', deletePlacement_placementName - The name of the empty placement to delete.

$sel:projectName:DeletePlacement', deletePlacement_projectName - The project containing the empty placement to delete.

data DeletePlacementResponse

See: newDeletePlacementResponse smart constructor.

Constructors

DeletePlacementResponse' Int

newDeletePlacementResponse

Arguments

:: Int
$sel:httpStatus:DeletePlacementResponse'

-> DeletePlacementResponse

Create a value of DeletePlacementResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:httpStatus:DeletePlacementResponse', deletePlacementResponse_httpStatus - The response's http status code.

UpdatePlacement

data UpdatePlacement

See: newUpdatePlacement smart constructor.

Constructors

UpdatePlacement' (Maybe (HashMap Text Text)) Text Text

newUpdatePlacement

Arguments

:: Text
$sel:placementName:UpdatePlacement'

-> Text
$sel:projectName:UpdatePlacement'

-> UpdatePlacement

Create a value of UpdatePlacement with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:attributes:UpdatePlacement', updatePlacement_attributes - The user-defined object of attributes used to update the placement. The maximum number of key/value pairs is 50.

$sel:placementName:UpdatePlacement', updatePlacement_placementName - The name of the placement to update.

$sel:projectName:UpdatePlacement', updatePlacement_projectName - The name of the project containing the placement to be updated.

data UpdatePlacementResponse

See: newUpdatePlacementResponse smart constructor.

Constructors

UpdatePlacementResponse' Int

newUpdatePlacementResponse

Arguments

:: Int
$sel:httpStatus:UpdatePlacementResponse'

-> UpdatePlacementResponse

Create a value of UpdatePlacementResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:httpStatus:UpdatePlacementResponse', updatePlacementResponse_httpStatus - The response's http status code.

TagResource

data TagResource

See: newTagResource smart constructor.

Constructors

TagResource' Text (HashMap Text Text)

newTagResource

Arguments

:: Text
$sel:resourceArn:TagResource'

-> TagResource

Create a value of TagResource with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:resourceArn:TagResource', tagResource_resourceArn - The ARN of the resouce for which tag(s) should be added or modified.

$sel:tags:TagResource', tagResource_tags - The new or modifying tag(s) for the resource. See AWS IoT 1-Click Service Limits for the maximum number of tags allowed per resource.

data TagResourceResponse

See: newTagResourceResponse smart constructor.

Constructors

TagResourceResponse' Int

newTagResourceResponse

Arguments

:: Int
$sel:httpStatus:TagResourceResponse'

-> TagResourceResponse

Create a value of TagResourceResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:httpStatus:TagResourceResponse', tagResourceResponse_httpStatus - The response's http status code.

ListPlacements (Paginated)

data ListPlacements

See: newListPlacements smart constructor.

Constructors

ListPlacements' (Maybe Text) (Maybe Natural) Text

newListPlacements

Arguments

:: Text
$sel:projectName:ListPlacements'

-> ListPlacements

Create a value of ListPlacements with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:nextToken:ListPlacements', listPlacements_nextToken - The token to retrieve the next set of results.

$sel:maxResults:ListPlacements', listPlacements_maxResults - The maximum number of results to return per request. If not set, a default value of 100 is used.

$sel:projectName:ListPlacements', listPlacements_projectName - The project containing the placements to be listed.

data ListPlacementsResponse

See: newListPlacementsResponse smart constructor.

Constructors

ListPlacementsResponse' (Maybe Text) Int [PlacementSummary]

newListPlacementsResponse

Arguments

:: Int
$sel:httpStatus:ListPlacementsResponse'

-> ListPlacementsResponse

Create a value of ListPlacementsResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:nextToken:ListPlacements', listPlacementsResponse_nextToken - The token used to retrieve the next set of results - will be effectively empty if there are no further results.

$sel:httpStatus:ListPlacementsResponse', listPlacementsResponse_httpStatus - The response's http status code.

$sel:placements:ListPlacementsResponse', listPlacementsResponse_placements - An object listing the requested placements.

UntagResource

data UntagResource

See: newUntagResource smart constructor.

Constructors

UntagResource' Text (NonEmpty Text)

newUntagResource

Arguments

:: Text
$sel:resourceArn:UntagResource'

-> NonEmpty Text
$sel:tagKeys:UntagResource'

-> UntagResource

Create a value of UntagResource with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:resourceArn:UntagResource', untagResource_resourceArn - The ARN of the resource whose tag you want to remove.

$sel:tagKeys:UntagResource', untagResource_tagKeys - The keys of those tags which you want to remove.

data UntagResourceResponse

See: newUntagResourceResponse smart constructor.

Constructors

UntagResourceResponse' Int

newUntagResourceResponse

Arguments

:: Int
$sel:httpStatus:UntagResourceResponse'

-> UntagResourceResponse

Create a value of UntagResourceResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:httpStatus:UntagResourceResponse', untagResourceResponse_httpStatus - The response's http status code.

CreateProject

data CreateProject

See: newCreateProject smart constructor.

Constructors

CreateProject' (Maybe PlacementTemplate) (Maybe Text) (Maybe (HashMap Text Text)) Text

newCreateProject

Arguments

:: Text
$sel:projectName:CreateProject'

-> CreateProject

Create a value of CreateProject with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:placementTemplate:CreateProject', createProject_placementTemplate - The schema defining the placement to be created. A placement template defines placement default attributes and device templates. You cannot add or remove device templates after the project has been created. However, you can update callbackOverrides for the device templates using the UpdateProject API.

$sel:description:CreateProject', createProject_description - An optional description for the project.

$sel:tags:CreateProject', createProject_tags - Optional tags (metadata key/value pairs) to be associated with the project. For example, { {"key1": "value1", "key2": "value2"} }. For more information, see AWS Tagging Strategies.

$sel:projectName:CreateProject', createProject_projectName - The name of the project to create.

data CreateProjectResponse

See: newCreateProjectResponse smart constructor.

Constructors

CreateProjectResponse' Int

newCreateProjectResponse

Arguments

:: Int
$sel:httpStatus:CreateProjectResponse'

-> CreateProjectResponse

Create a value of CreateProjectResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:httpStatus:CreateProjectResponse', createProjectResponse_httpStatus - The response's http status code.

Types

DeviceTemplate

data DeviceTemplate

An object representing a device for a placement template (see PlacementTemplate).

See: newDeviceTemplate smart constructor.

Constructors

DeviceTemplate' (Maybe Text) (Maybe (HashMap Text Text))

newDeviceTemplate :: DeviceTemplate

Create a value of DeviceTemplate with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:deviceType:DeviceTemplate', deviceTemplate_deviceType - The device type, which currently must be "button".

$sel:callbackOverrides:DeviceTemplate', deviceTemplate_callbackOverrides - An optional Lambda function to invoke instead of the default Lambda function provided by the placement template.

PlacementDescription

data PlacementDescription

An object describing a project's placement.

See: newPlacementDescription smart constructor.

Constructors

PlacementDescription' Text Text (HashMap Text Text) POSIX POSIX

newPlacementDescription

Arguments

:: Text
$sel:projectName:PlacementDescription'

-> Text
$sel:placementName:PlacementDescription'

-> UTCTime
$sel:createdDate:PlacementDescription'

-> UTCTime
$sel:updatedDate:PlacementDescription'

-> PlacementDescription

Create a value of PlacementDescription with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:projectName:PlacementDescription', placementDescription_projectName - The name of the project containing the placement.

$sel:placementName:PlacementDescription', placementDescription_placementName - The name of the placement.

$sel:attributes:PlacementDescription', placementDescription_attributes - The user-defined attributes associated with the placement.

$sel:createdDate:PlacementDescription', placementDescription_createdDate - The date when the placement was initially created, in UNIX epoch time format.

$sel:updatedDate:PlacementDescription', placementDescription_updatedDate - The date when the placement was last updated, in UNIX epoch time format. If the placement was not updated, then createdDate and updatedDate are the same.

PlacementSummary

data PlacementSummary

An object providing summary information for a particular placement.

See: newPlacementSummary smart constructor.

Constructors

PlacementSummary' Text Text POSIX POSIX

newPlacementSummary

Arguments

:: Text
$sel:projectName:PlacementSummary'

-> Text
$sel:placementName:PlacementSummary'

-> UTCTime
$sel:createdDate:PlacementSummary'

-> UTCTime
$sel:updatedDate:PlacementSummary'

-> PlacementSummary

Create a value of PlacementSummary with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:projectName:PlacementSummary', placementSummary_projectName - The name of the project containing the placement.

$sel:placementName:PlacementSummary', placementSummary_placementName - The name of the placement being summarized.

$sel:createdDate:PlacementSummary', placementSummary_createdDate - The date when the placement was originally created, in UNIX epoch time format.

$sel:updatedDate:PlacementSummary', placementSummary_updatedDate - The date when the placement was last updated, in UNIX epoch time format. If the placement was not updated, then createdDate and updatedDate are the same.

PlacementTemplate

data PlacementTemplate

An object defining the template for a placement.

See: newPlacementTemplate smart constructor.

Constructors

PlacementTemplate' (Maybe (HashMap Text DeviceTemplate)) (Maybe (HashMap Text Text))

newPlacementTemplate :: PlacementTemplate

Create a value of PlacementTemplate with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:deviceTemplates:PlacementTemplate', placementTemplate_deviceTemplates - An object specifying the DeviceTemplate for all placements using this (PlacementTemplate) template.

$sel:defaultAttributes:PlacementTemplate', placementTemplate_defaultAttributes - The default attributes (key/value pairs) to be applied to all placements using this template.

ProjectDescription

data ProjectDescription

An object providing detailed information for a particular project associated with an AWS account and region.

See: newProjectDescription smart constructor.

Constructors

ProjectDescription' (Maybe Text) (Maybe PlacementTemplate) (Maybe Text) (Maybe (HashMap Text Text)) Text POSIX POSIX

newProjectDescription

Arguments

:: Text
$sel:projectName:ProjectDescription'

-> UTCTime
$sel:createdDate:ProjectDescription'

-> UTCTime
$sel:updatedDate:ProjectDescription'

-> ProjectDescription

Create a value of ProjectDescription with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:arn:ProjectDescription', projectDescription_arn - The ARN of the project.

$sel:placementTemplate:ProjectDescription', projectDescription_placementTemplate - An object describing the project's placement specifications.

$sel:description:ProjectDescription', projectDescription_description - The description of the project.

$sel:tags:ProjectDescription', projectDescription_tags - The tags (metadata key/value pairs) associated with the project.

$sel:projectName:ProjectDescription', projectDescription_projectName - The name of the project for which to obtain information from.

$sel:createdDate:ProjectDescription', projectDescription_createdDate - The date when the project was originally created, in UNIX epoch time format.

$sel:updatedDate:ProjectDescription', projectDescription_updatedDate - The date when the project was last updated, in UNIX epoch time format. If the project was not updated, then createdDate and updatedDate are the same.

ProjectSummary

data ProjectSummary

An object providing summary information for a particular project for an associated AWS account and region.

See: newProjectSummary smart constructor.

Constructors

ProjectSummary' (Maybe Text) (Maybe (HashMap Text Text)) Text POSIX POSIX

newProjectSummary

Arguments

:: Text
$sel:projectName:ProjectSummary'

-> UTCTime
$sel:createdDate:ProjectSummary'

-> UTCTime
$sel:updatedDate:ProjectSummary'

-> ProjectSummary

Create a value of ProjectSummary with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:arn:ProjectSummary', projectSummary_arn - The ARN of the project.

$sel:tags:ProjectSummary', projectSummary_tags - The tags (metadata key/value pairs) associated with the project.

$sel:projectName:ProjectSummary', projectSummary_projectName - The name of the project being summarized.

$sel:createdDate:ProjectSummary', projectSummary_createdDate - The date when the project was originally created, in UNIX epoch time format.

$sel:updatedDate:ProjectSummary', projectSummary_updatedDate - The date when the project was last updated, in UNIX epoch time format. If the project was not updated, then createdDate and updatedDate are the same.