Amazonka.IoT1ClickProjects.DescribePlacement

Description

Describes a placement in a project.

Creating a Request

data DescribePlacement

See: newDescribePlacement smart constructor.

Constructors

DescribePlacement'

Fields

placementName :: Text

The name of the placement within a project.

projectName :: Text

The project containing the placement to be described.

newDescribePlacement

Arguments

:: Text
$sel:placementName:DescribePlacement'

-> Text
$sel:projectName:DescribePlacement'

-> DescribePlacement

Create a value of DescribePlacement with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:placementName:DescribePlacement', describePlacement_placementName - The name of the placement within a project.

$sel:projectName:DescribePlacement', describePlacement_projectName - The project containing the placement to be described.

Request Lenses

describePlacement_placementName :: Lens' DescribePlacement Text

The name of the placement within a project.

describePlacement_projectName :: Lens' DescribePlacement Text

The project containing the placement to be described.

Destructuring the Response

data DescribePlacementResponse

See: newDescribePlacementResponse smart constructor.

Constructors

DescribePlacementResponse'

Fields

httpStatus :: Int

The response's http status code.

placement :: PlacementDescription

An object describing the placement.

newDescribePlacementResponse

Arguments

:: Int
$sel:httpStatus:DescribePlacementResponse'

-> PlacementDescription
$sel:placement:DescribePlacementResponse'

-> DescribePlacementResponse

Create a value of DescribePlacementResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:httpStatus:DescribePlacementResponse', describePlacementResponse_httpStatus - The response's http status code.

$sel:placement:DescribePlacementResponse', describePlacementResponse_placement - An object describing the placement.

Response Lenses

describePlacementResponse_httpStatus :: Lens' DescribePlacementResponse Int

The response's http status code.

describePlacementResponse_placement :: Lens' DescribePlacementResponse PlacementDescription

An object describing the placement.