Amazonka.IoT1ClickProjects.Lens

Description

Operations

ListProjects

listProjects_nextToken :: Lens' ListProjects (Maybe Text)

The token to retrieve the next set of results.

listProjects_maxResults :: Lens' ListProjects (Maybe Natural)

The maximum number of results to return per request. If not set, a default value of 100 is used.

listProjectsResponse_nextToken :: Lens' ListProjectsResponse (Maybe Text)

The token used to retrieve the next set of results - will be effectively empty if there are no further results.

listProjectsResponse_httpStatus :: Lens' ListProjectsResponse Int

The response's http status code.

listProjectsResponse_projects :: Lens' ListProjectsResponse [ProjectSummary]

An object containing the list of projects.

DeleteProject

deleteProject_projectName :: Lens' DeleteProject Text

The name of the empty project to delete.

deleteProjectResponse_httpStatus :: Lens' DeleteProjectResponse Int

The response's http status code.

UpdateProject

updateProject_placementTemplate :: Lens' UpdateProject (Maybe PlacementTemplate)

An object defining the project update. Once a project has been created, you cannot add device template names to the project. However, for a given placementTemplate, you can update the associated callbackOverrides for the device definition using this API.

updateProject_description :: Lens' UpdateProject (Maybe Text)

An optional user-defined description for the project.

updateProject_projectName :: Lens' UpdateProject Text

The name of the project to be updated.

updateProjectResponse_httpStatus :: Lens' UpdateProjectResponse Int

The response's http status code.

ListTagsForResource

listTagsForResource_resourceArn :: Lens' ListTagsForResource Text

The ARN of the resource whose tags you want to list.

listTagsForResourceResponse_tags :: Lens' ListTagsForResourceResponse (Maybe (HashMap Text Text))

The tags (metadata key/value pairs) which you have assigned to the resource.

listTagsForResourceResponse_httpStatus :: Lens' ListTagsForResourceResponse Int

The response's http status code.

DisassociateDeviceFromPlacement

disassociateDeviceFromPlacement_projectName :: Lens' DisassociateDeviceFromPlacement Text

The name of the project that contains the placement.

disassociateDeviceFromPlacement_placementName :: Lens' DisassociateDeviceFromPlacement Text

The name of the placement that the device should be removed from.

disassociateDeviceFromPlacement_deviceTemplateName :: Lens' DisassociateDeviceFromPlacement Text

The device ID that should be removed from the placement.

disassociateDeviceFromPlacementResponse_httpStatus :: Lens' DisassociateDeviceFromPlacementResponse Int

The response's http status code.

CreatePlacement

createPlacement_attributes :: Lens' CreatePlacement (Maybe (HashMap Text Text))

Optional user-defined key/value pairs providing contextual data (such as location or function) for the placement.

createPlacement_placementName :: Lens' CreatePlacement Text

The name of the placement to be created.

createPlacement_projectName :: Lens' CreatePlacement Text

The name of the project in which to create the placement.

createPlacementResponse_httpStatus :: Lens' CreatePlacementResponse Int

The response's http status code.

DescribeProject

describeProject_projectName :: Lens' DescribeProject Text

The name of the project to be described.

describeProjectResponse_httpStatus :: Lens' DescribeProjectResponse Int

The response's http status code.

describeProjectResponse_project :: Lens' DescribeProjectResponse ProjectDescription

An object describing the project.

GetDevicesInPlacement

getDevicesInPlacement_projectName :: Lens' GetDevicesInPlacement Text

The name of the project containing the placement.

getDevicesInPlacement_placementName :: Lens' GetDevicesInPlacement Text

The name of the placement to get the devices from.

getDevicesInPlacementResponse_httpStatus :: Lens' GetDevicesInPlacementResponse Int

The response's http status code.

getDevicesInPlacementResponse_devices :: Lens' GetDevicesInPlacementResponse (HashMap Text Text)

An object containing the devices (zero or more) within the placement.

DescribePlacement

describePlacement_placementName :: Lens' DescribePlacement Text

The name of the placement within a project.

describePlacement_projectName :: Lens' DescribePlacement Text

The project containing the placement to be described.

describePlacementResponse_httpStatus :: Lens' DescribePlacementResponse Int

The response's http status code.

describePlacementResponse_placement :: Lens' DescribePlacementResponse PlacementDescription

An object describing the placement.

AssociateDeviceWithPlacement

associateDeviceWithPlacement_projectName :: Lens' AssociateDeviceWithPlacement Text

The name of the project containing the placement in which to associate the device.

associateDeviceWithPlacement_placementName :: Lens' AssociateDeviceWithPlacement Text

The name of the placement in which to associate the device.

associateDeviceWithPlacement_deviceId :: Lens' AssociateDeviceWithPlacement Text

The ID of the physical device to be associated with the given placement in the project. Note that a mandatory 4 character prefix is required for all deviceId values.

associateDeviceWithPlacement_deviceTemplateName :: Lens' AssociateDeviceWithPlacement Text

The device template name to associate with the device ID.

associateDeviceWithPlacementResponse_httpStatus :: Lens' AssociateDeviceWithPlacementResponse Int

The response's http status code.

DeletePlacement

deletePlacement_placementName :: Lens' DeletePlacement Text

The name of the empty placement to delete.

deletePlacement_projectName :: Lens' DeletePlacement Text

The project containing the empty placement to delete.

deletePlacementResponse_httpStatus :: Lens' DeletePlacementResponse Int

The response's http status code.

UpdatePlacement

updatePlacement_attributes :: Lens' UpdatePlacement (Maybe (HashMap Text Text))

The user-defined object of attributes used to update the placement. The maximum number of key/value pairs is 50.

updatePlacement_placementName :: Lens' UpdatePlacement Text

The name of the placement to update.

updatePlacement_projectName :: Lens' UpdatePlacement Text

The name of the project containing the placement to be updated.

updatePlacementResponse_httpStatus :: Lens' UpdatePlacementResponse Int

The response's http status code.

TagResource

tagResource_resourceArn :: Lens' TagResource Text

The ARN of the resouce for which tag(s) should be added or modified.

tagResource_tags :: Lens' TagResource (HashMap Text Text)

The new or modifying tag(s) for the resource. See AWS IoT 1-Click Service Limits for the maximum number of tags allowed per resource.

tagResourceResponse_httpStatus :: Lens' TagResourceResponse Int

The response's http status code.

ListPlacements

listPlacements_nextToken :: Lens' ListPlacements (Maybe Text)

The token to retrieve the next set of results.

listPlacements_maxResults :: Lens' ListPlacements (Maybe Natural)

The maximum number of results to return per request. If not set, a default value of 100 is used.

listPlacements_projectName :: Lens' ListPlacements Text

The project containing the placements to be listed.

listPlacementsResponse_nextToken :: Lens' ListPlacementsResponse (Maybe Text)

The token used to retrieve the next set of results - will be effectively empty if there are no further results.

listPlacementsResponse_httpStatus :: Lens' ListPlacementsResponse Int

The response's http status code.

listPlacementsResponse_placements :: Lens' ListPlacementsResponse [PlacementSummary]

An object listing the requested placements.

UntagResource

untagResource_resourceArn :: Lens' UntagResource Text

The ARN of the resource whose tag you want to remove.

untagResource_tagKeys :: Lens' UntagResource (NonEmpty Text)

The keys of those tags which you want to remove.

untagResourceResponse_httpStatus :: Lens' UntagResourceResponse Int

The response's http status code.

CreateProject

createProject_placementTemplate :: Lens' CreateProject (Maybe PlacementTemplate)

The schema defining the placement to be created. A placement template defines placement default attributes and device templates. You cannot add or remove device templates after the project has been created. However, you can update callbackOverrides for the device templates using the UpdateProject API.

createProject_description :: Lens' CreateProject (Maybe Text)

An optional description for the project.

createProject_tags :: Lens' CreateProject (Maybe (HashMap Text Text))

Optional tags (metadata key/value pairs) to be associated with the project. For example, { {"key1": "value1", "key2": "value2"} }. For more information, see AWS Tagging Strategies.

createProject_projectName :: Lens' CreateProject Text

The name of the project to create.

createProjectResponse_httpStatus :: Lens' CreateProjectResponse Int

The response's http status code.

Types

DeviceTemplate

deviceTemplate_deviceType :: Lens' DeviceTemplate (Maybe Text)

The device type, which currently must be "button".

deviceTemplate_callbackOverrides :: Lens' DeviceTemplate (Maybe (HashMap Text Text))

An optional Lambda function to invoke instead of the default Lambda function provided by the placement template.

PlacementDescription

placementDescription_projectName :: Lens' PlacementDescription Text

The name of the project containing the placement.

placementDescription_placementName :: Lens' PlacementDescription Text

The name of the placement.

placementDescription_attributes :: Lens' PlacementDescription (HashMap Text Text)

The user-defined attributes associated with the placement.

placementDescription_createdDate :: Lens' PlacementDescription UTCTime

The date when the placement was initially created, in UNIX epoch time format.

placementDescription_updatedDate :: Lens' PlacementDescription UTCTime

The date when the placement was last updated, in UNIX epoch time format. If the placement was not updated, then createdDate and updatedDate are the same.

PlacementSummary

placementSummary_projectName :: Lens' PlacementSummary Text

The name of the project containing the placement.

placementSummary_placementName :: Lens' PlacementSummary Text

The name of the placement being summarized.

placementSummary_createdDate :: Lens' PlacementSummary UTCTime

The date when the placement was originally created, in UNIX epoch time format.

placementSummary_updatedDate :: Lens' PlacementSummary UTCTime

The date when the placement was last updated, in UNIX epoch time format. If the placement was not updated, then createdDate and updatedDate are the same.

PlacementTemplate

placementTemplate_deviceTemplates :: Lens' PlacementTemplate (Maybe (HashMap Text DeviceTemplate))

An object specifying the DeviceTemplate for all placements using this (PlacementTemplate) template.

placementTemplate_defaultAttributes :: Lens' PlacementTemplate (Maybe (HashMap Text Text))

The default attributes (key/value pairs) to be applied to all placements using this template.

ProjectDescription

projectDescription_arn :: Lens' ProjectDescription (Maybe Text)

The ARN of the project.

projectDescription_placementTemplate :: Lens' ProjectDescription (Maybe PlacementTemplate)

An object describing the project's placement specifications.

projectDescription_description :: Lens' ProjectDescription (Maybe Text)

The description of the project.

projectDescription_tags :: Lens' ProjectDescription (Maybe (HashMap Text Text))

The tags (metadata key/value pairs) associated with the project.

projectDescription_projectName :: Lens' ProjectDescription Text

The name of the project for which to obtain information from.

projectDescription_createdDate :: Lens' ProjectDescription UTCTime

The date when the project was originally created, in UNIX epoch time format.

projectDescription_updatedDate :: Lens' ProjectDescription UTCTime

The date when the project was last updated, in UNIX epoch time format. If the project was not updated, then createdDate and updatedDate are the same.

ProjectSummary

projectSummary_arn :: Lens' ProjectSummary (Maybe Text)

The ARN of the project.

projectSummary_tags :: Lens' ProjectSummary (Maybe (HashMap Text Text))

The tags (metadata key/value pairs) associated with the project.

projectSummary_projectName :: Lens' ProjectSummary Text

The name of the project being summarized.

projectSummary_createdDate :: Lens' ProjectSummary UTCTime

The date when the project was originally created, in UNIX epoch time format.

projectSummary_updatedDate :: Lens' ProjectSummary UTCTime

The date when the project was last updated, in UNIX epoch time format. If the project was not updated, then createdDate and updatedDate are the same.