Amazonka.IoT1ClickProjects.Types.PlacementDescription

Description

Documentation

data PlacementDescription

An object describing a project's placement.

See: newPlacementDescription smart constructor.

Constructors

PlacementDescription'

Fields

projectName :: Text

The name of the project containing the placement.

placementName :: Text

The name of the placement.

attributes :: HashMap Text Text

The user-defined attributes associated with the placement.

createdDate :: POSIX

The date when the placement was initially created, in UNIX epoch time format.

updatedDate :: POSIX

The date when the placement was last updated, in UNIX epoch time format. If the placement was not updated, then createdDate and updatedDate are the same.

newPlacementDescription

Arguments

:: Text
$sel:projectName:PlacementDescription'

-> Text
$sel:placementName:PlacementDescription'

-> UTCTime
$sel:createdDate:PlacementDescription'

-> UTCTime
$sel:updatedDate:PlacementDescription'

-> PlacementDescription

Create a value of PlacementDescription with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:projectName:PlacementDescription', placementDescription_projectName - The name of the project containing the placement.

$sel:placementName:PlacementDescription', placementDescription_placementName - The name of the placement.

$sel:attributes:PlacementDescription', placementDescription_attributes - The user-defined attributes associated with the placement.

$sel:createdDate:PlacementDescription', placementDescription_createdDate - The date when the placement was initially created, in UNIX epoch time format.

$sel:updatedDate:PlacementDescription', placementDescription_updatedDate - The date when the placement was last updated, in UNIX epoch time format. If the placement was not updated, then createdDate and updatedDate are the same.

placementDescription_projectName :: Lens' PlacementDescription Text

The name of the project containing the placement.

placementDescription_placementName :: Lens' PlacementDescription Text

The name of the placement.

placementDescription_attributes :: Lens' PlacementDescription (HashMap Text Text)

The user-defined attributes associated with the placement.

placementDescription_createdDate :: Lens' PlacementDescription UTCTime

The date when the placement was initially created, in UNIX epoch time format.

placementDescription_updatedDate :: Lens' PlacementDescription UTCTime

The date when the placement was last updated, in UNIX epoch time format. If the placement was not updated, then createdDate and updatedDate are the same.