Amazonka.IoTAnalytics.ListDatasetContents

Description

Lists information about dataset contents that have been created.

This operation returns paginated results.

Creating a Request

data ListDatasetContents

See: newListDatasetContents smart constructor.

Constructors

ListDatasetContents'

Fields

nextToken :: Maybe Text

The token for the next set of results.

scheduledBefore :: Maybe POSIX

A filter to limit results to those dataset contents whose creation is scheduled before the given time. See the field triggers.schedule in the CreateDataset request. (timestamp)

maxResults :: Maybe Natural

The maximum number of results to return in this request.

scheduledOnOrAfter :: Maybe POSIX

A filter to limit results to those dataset contents whose creation is scheduled on or after the given time. See the field triggers.schedule in the CreateDataset request. (timestamp)

datasetName :: Text

The name of the dataset whose contents information you want to list.

newListDatasetContents

Arguments

:: Text
$sel:datasetName:ListDatasetContents'

-> ListDatasetContents

Create a value of ListDatasetContents with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:nextToken:ListDatasetContents', listDatasetContents_nextToken - The token for the next set of results.

$sel:scheduledBefore:ListDatasetContents', listDatasetContents_scheduledBefore - A filter to limit results to those dataset contents whose creation is scheduled before the given time. See the field triggers.schedule in the CreateDataset request. (timestamp)

$sel:maxResults:ListDatasetContents', listDatasetContents_maxResults - The maximum number of results to return in this request.

$sel:scheduledOnOrAfter:ListDatasetContents', listDatasetContents_scheduledOnOrAfter - A filter to limit results to those dataset contents whose creation is scheduled on or after the given time. See the field triggers.schedule in the CreateDataset request. (timestamp)

$sel:datasetName:ListDatasetContents', listDatasetContents_datasetName - The name of the dataset whose contents information you want to list.

Request Lenses

listDatasetContents_nextToken :: Lens' ListDatasetContents (Maybe Text)

The token for the next set of results.

listDatasetContents_scheduledBefore :: Lens' ListDatasetContents (Maybe UTCTime)

A filter to limit results to those dataset contents whose creation is scheduled before the given time. See the field triggers.schedule in the CreateDataset request. (timestamp)

listDatasetContents_maxResults :: Lens' ListDatasetContents (Maybe Natural)

The maximum number of results to return in this request.

listDatasetContents_scheduledOnOrAfter :: Lens' ListDatasetContents (Maybe UTCTime)

A filter to limit results to those dataset contents whose creation is scheduled on or after the given time. See the field triggers.schedule in the CreateDataset request. (timestamp)

listDatasetContents_datasetName :: Lens' ListDatasetContents Text

The name of the dataset whose contents information you want to list.

Destructuring the Response

data ListDatasetContentsResponse

See: newListDatasetContentsResponse smart constructor.

Constructors

ListDatasetContentsResponse'

Fields

datasetContentSummaries :: Maybe [DatasetContentSummary]

Summary information about dataset contents that have been created.

nextToken :: Maybe Text

The token to retrieve the next set of results, or null if there are no more results.

httpStatus :: Int

The response's http status code.

newListDatasetContentsResponse

Arguments

:: Int
$sel:httpStatus:ListDatasetContentsResponse'

-> ListDatasetContentsResponse

Create a value of ListDatasetContentsResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:datasetContentSummaries:ListDatasetContentsResponse', listDatasetContentsResponse_datasetContentSummaries - Summary information about dataset contents that have been created.

$sel:nextToken:ListDatasetContents', listDatasetContentsResponse_nextToken - The token to retrieve the next set of results, or null if there are no more results.

$sel:httpStatus:ListDatasetContentsResponse', listDatasetContentsResponse_httpStatus - The response's http status code.

Response Lenses

listDatasetContentsResponse_datasetContentSummaries :: Lens' ListDatasetContentsResponse (Maybe [DatasetContentSummary])

Summary information about dataset contents that have been created.

listDatasetContentsResponse_nextToken :: Lens' ListDatasetContentsResponse (Maybe Text)

The token to retrieve the next set of results, or null if there are no more results.

listDatasetContentsResponse_httpStatus :: Lens' ListDatasetContentsResponse Int

The response's http status code.