Amazonka.IoTAnalytics.SampleChannelData

Description

Retrieves a sample of messages from the specified channel ingested during the specified timeframe. Up to 10 messages can be retrieved.

Creating a Request

data SampleChannelData

See: newSampleChannelData smart constructor.

Constructors

SampleChannelData'

Fields

startTime :: Maybe POSIX

The start of the time window from which sample messages are retrieved.

maxMessages :: Maybe Natural

The number of sample messages to be retrieved. The limit is 10. The default is also 10.

endTime :: Maybe POSIX

The end of the time window from which sample messages are retrieved.

channelName :: Text

The name of the channel whose message samples are retrieved.

newSampleChannelData

Arguments

:: Text
$sel:channelName:SampleChannelData'

-> SampleChannelData

Create a value of SampleChannelData with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:startTime:SampleChannelData', sampleChannelData_startTime - The start of the time window from which sample messages are retrieved.

$sel:maxMessages:SampleChannelData', sampleChannelData_maxMessages - The number of sample messages to be retrieved. The limit is 10. The default is also 10.

$sel:endTime:SampleChannelData', sampleChannelData_endTime - The end of the time window from which sample messages are retrieved.

$sel:channelName:SampleChannelData', sampleChannelData_channelName - The name of the channel whose message samples are retrieved.

Request Lenses

sampleChannelData_startTime :: Lens' SampleChannelData (Maybe UTCTime)

The start of the time window from which sample messages are retrieved.

sampleChannelData_maxMessages :: Lens' SampleChannelData (Maybe Natural)

The number of sample messages to be retrieved. The limit is 10. The default is also 10.

sampleChannelData_endTime :: Lens' SampleChannelData (Maybe UTCTime)

The end of the time window from which sample messages are retrieved.

sampleChannelData_channelName :: Lens' SampleChannelData Text

The name of the channel whose message samples are retrieved.

Destructuring the Response

data SampleChannelDataResponse

See: newSampleChannelDataResponse smart constructor.

Constructors

SampleChannelDataResponse'

Fields

payloads :: Maybe (NonEmpty Base64)

The list of message samples. Each sample message is returned as a base64-encoded string.

httpStatus :: Int

The response's http status code.

newSampleChannelDataResponse

Arguments

:: Int
$sel:httpStatus:SampleChannelDataResponse'

-> SampleChannelDataResponse

Create a value of SampleChannelDataResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:payloads:SampleChannelDataResponse', sampleChannelDataResponse_payloads - The list of message samples. Each sample message is returned as a base64-encoded string.

$sel:httpStatus:SampleChannelDataResponse', sampleChannelDataResponse_httpStatus - The response's http status code.

Response Lenses

sampleChannelDataResponse_payloads :: Lens' SampleChannelDataResponse (Maybe (NonEmpty ByteString))

The list of message samples. Each sample message is returned as a base64-encoded string.

sampleChannelDataResponse_httpStatus :: Lens' SampleChannelDataResponse Int

The response's http status code.