Amazonka.IoTAnalytics.Types.Dataset

Description

Documentation

data Dataset

Information about a dataset.

See: newDataset smart constructor.

Constructors

Dataset'

Fields

creationTime :: Maybe POSIX

When the dataset was created.

status :: Maybe DatasetStatus

The status of the dataset.

versioningConfiguration :: Maybe VersioningConfiguration

Optional. How many versions of dataset contents are kept. If not specified or set to null, only the latest version plus the latest succeeded version (if they are different) are kept for the time period specified by the retentionPeriod parameter. For more information, see Keeping Multiple Versions of IoT Analytics datasets in the IoT Analytics User Guide.

arn :: Maybe Text

The ARN of the dataset.

actions :: Maybe (NonEmpty DatasetAction)

The DatasetAction objects that automatically create the dataset contents.

triggers :: Maybe [DatasetTrigger]

The DatasetTrigger objects that specify when the dataset is automatically updated.

retentionPeriod :: Maybe RetentionPeriod

Optional. How long, in days, message data is kept for the dataset.

lateDataRules :: Maybe (NonEmpty LateDataRule)

A list of data rules that send notifications to CloudWatch, when data arrives late. To specify lateDataRules, the dataset must use a DeltaTimer filter.

name :: Maybe Text

The name of the dataset.

contentDeliveryRules :: Maybe [DatasetContentDeliveryRule]

When dataset contents are created they are delivered to destinations specified here.

lastUpdateTime :: Maybe POSIX

The last time the dataset was updated.

newDataset :: Dataset

Create a value of Dataset with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:creationTime:Dataset', dataset_creationTime - When the dataset was created.

$sel:status:Dataset', dataset_status - The status of the dataset.

$sel:versioningConfiguration:Dataset', dataset_versioningConfiguration - Optional. How many versions of dataset contents are kept. If not specified or set to null, only the latest version plus the latest succeeded version (if they are different) are kept for the time period specified by the retentionPeriod parameter. For more information, see Keeping Multiple Versions of IoT Analytics datasets in the IoT Analytics User Guide.

$sel:arn:Dataset', dataset_arn - The ARN of the dataset.

$sel:actions:Dataset', dataset_actions - The DatasetAction objects that automatically create the dataset contents.

$sel:triggers:Dataset', dataset_triggers - The DatasetTrigger objects that specify when the dataset is automatically updated.

$sel:retentionPeriod:Dataset', dataset_retentionPeriod - Optional. How long, in days, message data is kept for the dataset.

$sel:lateDataRules:Dataset', dataset_lateDataRules - A list of data rules that send notifications to CloudWatch, when data arrives late. To specify lateDataRules, the dataset must use a DeltaTimer filter.

$sel:name:Dataset', dataset_name - The name of the dataset.

$sel:contentDeliveryRules:Dataset', dataset_contentDeliveryRules - When dataset contents are created they are delivered to destinations specified here.

$sel:lastUpdateTime:Dataset', dataset_lastUpdateTime - The last time the dataset was updated.

dataset_creationTime :: Lens' Dataset (Maybe UTCTime)

When the dataset was created.

dataset_status :: Lens' Dataset (Maybe DatasetStatus)

The status of the dataset.

dataset_versioningConfiguration :: Lens' Dataset (Maybe VersioningConfiguration)

Optional. How many versions of dataset contents are kept. If not specified or set to null, only the latest version plus the latest succeeded version (if they are different) are kept for the time period specified by the retentionPeriod parameter. For more information, see Keeping Multiple Versions of IoT Analytics datasets in the IoT Analytics User Guide.

dataset_arn :: Lens' Dataset (Maybe Text)

The ARN of the dataset.

dataset_actions :: Lens' Dataset (Maybe (NonEmpty DatasetAction))

The DatasetAction objects that automatically create the dataset contents.

dataset_triggers :: Lens' Dataset (Maybe [DatasetTrigger])

The DatasetTrigger objects that specify when the dataset is automatically updated.

dataset_retentionPeriod :: Lens' Dataset (Maybe RetentionPeriod)

Optional. How long, in days, message data is kept for the dataset.

dataset_lateDataRules :: Lens' Dataset (Maybe (NonEmpty LateDataRule))

A list of data rules that send notifications to CloudWatch, when data arrives late. To specify lateDataRules, the dataset must use a DeltaTimer filter.

dataset_name :: Lens' Dataset (Maybe Text)

The name of the dataset.

dataset_contentDeliveryRules :: Lens' Dataset (Maybe [DatasetContentDeliveryRule])

When dataset contents are created they are delivered to destinations specified here.

dataset_lastUpdateTime :: Lens' Dataset (Maybe UTCTime)

The last time the dataset was updated.