Amazonka.IoTAnalytics.Types.DatasetSummary

Description

Documentation

data DatasetSummary

A summary of information about a dataset.

See: newDatasetSummary smart constructor.

Constructors

DatasetSummary'

Fields

creationTime :: Maybe POSIX

The time the dataset was created.

status :: Maybe DatasetStatus

The status of the dataset.

actions :: Maybe (NonEmpty DatasetActionSummary)

A list of DataActionSummary objects.

triggers :: Maybe [DatasetTrigger]

A list of triggers. A trigger causes dataset content to be populated at a specified time interval or when another dataset is populated. The list of triggers can be empty or contain up to five DataSetTrigger objects

datasetName :: Maybe Text

The name of the dataset.

lastUpdateTime :: Maybe POSIX

The last time the dataset was updated.

newDatasetSummary :: DatasetSummary

Create a value of DatasetSummary with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:creationTime:DatasetSummary', datasetSummary_creationTime - The time the dataset was created.

$sel:status:DatasetSummary', datasetSummary_status - The status of the dataset.

$sel:actions:DatasetSummary', datasetSummary_actions - A list of DataActionSummary objects.

$sel:triggers:DatasetSummary', datasetSummary_triggers - A list of triggers. A trigger causes dataset content to be populated at a specified time interval or when another dataset is populated. The list of triggers can be empty or contain up to five DataSetTrigger objects

$sel:datasetName:DatasetSummary', datasetSummary_datasetName - The name of the dataset.

$sel:lastUpdateTime:DatasetSummary', datasetSummary_lastUpdateTime - The last time the dataset was updated.

datasetSummary_creationTime :: Lens' DatasetSummary (Maybe UTCTime)

The time the dataset was created.

datasetSummary_status :: Lens' DatasetSummary (Maybe DatasetStatus)

The status of the dataset.

datasetSummary_actions :: Lens' DatasetSummary (Maybe (NonEmpty DatasetActionSummary))

A list of DataActionSummary objects.

datasetSummary_triggers :: Lens' DatasetSummary (Maybe [DatasetTrigger])

A list of triggers. A trigger causes dataset content to be populated at a specified time interval or when another dataset is populated. The list of triggers can be empty or contain up to five DataSetTrigger objects

datasetSummary_datasetName :: Lens' DatasetSummary (Maybe Text)

The name of the dataset.

datasetSummary_lastUpdateTime :: Lens' DatasetSummary (Maybe UTCTime)

The last time the dataset was updated.