Amazonka.IoTAnalytics.Types.DatastoreSummary

Description

Documentation

data DatastoreSummary

A summary of information about a data store.

See: newDatastoreSummary smart constructor.

Constructors

DatastoreSummary'

Fields

creationTime :: Maybe POSIX

When the data store was created.

status :: Maybe DatastoreStatus

The status of the data store.

lastMessageArrivalTime :: Maybe POSIX

The last time when a new message arrived in the data store.

IoT Analytics updates this value at most once per minute for Amazon Simple Storage Service one data store. Hence, the lastMessageArrivalTime value is an approximation.

This feature only applies to messages that arrived in the data store after October 23, 2020.

datastoreName :: Maybe Text

The name of the data store.

datastorePartitions :: Maybe DatastorePartitions

Contains information about the partition dimensions in a data store.

fileFormatType :: Maybe FileFormatType

The file format of the data in the data store.

lastUpdateTime :: Maybe POSIX

The last time the data store was updated.

datastoreStorage :: Maybe DatastoreStorageSummary

Where data in a data store is stored.

newDatastoreSummary :: DatastoreSummary

Create a value of DatastoreSummary with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:creationTime:DatastoreSummary', datastoreSummary_creationTime - When the data store was created.

$sel:status:DatastoreSummary', datastoreSummary_status - The status of the data store.

$sel:lastMessageArrivalTime:DatastoreSummary', datastoreSummary_lastMessageArrivalTime - The last time when a new message arrived in the data store.

IoT Analytics updates this value at most once per minute for Amazon Simple Storage Service one data store. Hence, the lastMessageArrivalTime value is an approximation.

This feature only applies to messages that arrived in the data store after October 23, 2020.

$sel:datastoreName:DatastoreSummary', datastoreSummary_datastoreName - The name of the data store.

$sel:datastorePartitions:DatastoreSummary', datastoreSummary_datastorePartitions - Contains information about the partition dimensions in a data store.

$sel:fileFormatType:DatastoreSummary', datastoreSummary_fileFormatType - The file format of the data in the data store.

$sel:lastUpdateTime:DatastoreSummary', datastoreSummary_lastUpdateTime - The last time the data store was updated.

$sel:datastoreStorage:DatastoreSummary', datastoreSummary_datastoreStorage - Where data in a data store is stored.

datastoreSummary_creationTime :: Lens' DatastoreSummary (Maybe UTCTime)

When the data store was created.

datastoreSummary_status :: Lens' DatastoreSummary (Maybe DatastoreStatus)

The status of the data store.

datastoreSummary_lastMessageArrivalTime :: Lens' DatastoreSummary (Maybe UTCTime)

The last time when a new message arrived in the data store.

IoT Analytics updates this value at most once per minute for Amazon Simple Storage Service one data store. Hence, the lastMessageArrivalTime value is an approximation.

This feature only applies to messages that arrived in the data store after October 23, 2020.

datastoreSummary_datastoreName :: Lens' DatastoreSummary (Maybe Text)

The name of the data store.

datastoreSummary_datastorePartitions :: Lens' DatastoreSummary (Maybe DatastorePartitions)

Contains information about the partition dimensions in a data store.

datastoreSummary_fileFormatType :: Lens' DatastoreSummary (Maybe FileFormatType)

The file format of the data in the data store.

datastoreSummary_lastUpdateTime :: Lens' DatastoreSummary (Maybe UTCTime)

The last time the data store was updated.

datastoreSummary_datastoreStorage :: Lens' DatastoreSummary (Maybe DatastoreStorageSummary)

Where data in a data store is stored.