Amazonka.IoTAnalytics.Types.DeltaTime

Description

Documentation

data DeltaTime

Used to limit data to that which has arrived since the last execution of the action.

See: newDeltaTime smart constructor.

Constructors

DeltaTime'

Fields

offsetSeconds :: Int

The number of seconds of estimated in-flight lag time of message data. When you create dataset contents using message data from a specified timeframe, some message data might still be in flight when processing begins, and so do not arrive in time to be processed. Use this field to make allowances for the in flight time of your message data, so that data not processed from a previous timeframe is included with the next timeframe. Otherwise, missed message data would be excluded from processing during the next timeframe too, because its timestamp places it within the previous timeframe.

timeExpression :: Text

An expression by which the time of the message data might be determined. This can be the name of a timestamp field or a SQL expression that is used to derive the time the message data was generated.

newDeltaTime

Arguments

:: Int
$sel:offsetSeconds:DeltaTime'

-> Text
$sel:timeExpression:DeltaTime'

-> DeltaTime

Create a value of DeltaTime with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:offsetSeconds:DeltaTime', deltaTime_offsetSeconds - The number of seconds of estimated in-flight lag time of message data. When you create dataset contents using message data from a specified timeframe, some message data might still be in flight when processing begins, and so do not arrive in time to be processed. Use this field to make allowances for the in flight time of your message data, so that data not processed from a previous timeframe is included with the next timeframe. Otherwise, missed message data would be excluded from processing during the next timeframe too, because its timestamp places it within the previous timeframe.

$sel:timeExpression:DeltaTime', deltaTime_timeExpression - An expression by which the time of the message data might be determined. This can be the name of a timestamp field or a SQL expression that is used to derive the time the message data was generated.

deltaTime_offsetSeconds :: Lens' DeltaTime Int

The number of seconds of estimated in-flight lag time of message data. When you create dataset contents using message data from a specified timeframe, some message data might still be in flight when processing begins, and so do not arrive in time to be processed. Use this field to make allowances for the in flight time of your message data, so that data not processed from a previous timeframe is included with the next timeframe. Otherwise, missed message data would be excluded from processing during the next timeframe too, because its timestamp places it within the previous timeframe.

deltaTime_timeExpression :: Lens' DeltaTime Text

An expression by which the time of the message data might be determined. This can be the name of a timestamp field or a SQL expression that is used to derive the time the message data was generated.