Amazonka.IoTAnalytics.Types.PipelineActivity

Description

Documentation

data PipelineActivity

An activity that performs a transformation on a message.

See: newPipelineActivity smart constructor.

Constructors

PipelineActivity'

Fields

selectAttributes :: Maybe SelectAttributesActivity

Used to create a new message using only the specified attributes from the original message.

channel :: Maybe ChannelActivity

Determines the source of the messages to be processed.

addAttributes :: Maybe AddAttributesActivity

Adds other attributes based on existing attributes in the message.

deviceRegistryEnrich :: Maybe DeviceRegistryEnrichActivity

Adds data from the IoT device registry to your message.

removeAttributes :: Maybe RemoveAttributesActivity

Removes attributes from a message.

lambda :: Maybe LambdaActivity

Runs a Lambda function to modify the message.

datastore :: Maybe DatastoreActivity

Specifies where to store the processed message data.

deviceShadowEnrich :: Maybe DeviceShadowEnrichActivity

Adds information from the IoT Device Shadow service to a message.

filter' :: Maybe FilterActivity

Filters a message based on its attributes.

math :: Maybe MathActivity

Computes an arithmetic expression using the message's attributes and adds it to the message.

newPipelineActivity :: PipelineActivity

Create a value of PipelineActivity with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:selectAttributes:PipelineActivity', pipelineActivity_selectAttributes - Used to create a new message using only the specified attributes from the original message.

$sel:channel:PipelineActivity', pipelineActivity_channel - Determines the source of the messages to be processed.

$sel:addAttributes:PipelineActivity', pipelineActivity_addAttributes - Adds other attributes based on existing attributes in the message.

$sel:deviceRegistryEnrich:PipelineActivity', pipelineActivity_deviceRegistryEnrich - Adds data from the IoT device registry to your message.

$sel:removeAttributes:PipelineActivity', pipelineActivity_removeAttributes - Removes attributes from a message.

$sel:lambda:PipelineActivity', pipelineActivity_lambda - Runs a Lambda function to modify the message.

$sel:datastore:PipelineActivity', pipelineActivity_datastore - Specifies where to store the processed message data.

$sel:deviceShadowEnrich:PipelineActivity', pipelineActivity_deviceShadowEnrich - Adds information from the IoT Device Shadow service to a message.

$sel:filter':PipelineActivity', pipelineActivity_filter - Filters a message based on its attributes.

$sel:math:PipelineActivity', pipelineActivity_math - Computes an arithmetic expression using the message's attributes and adds it to the message.

pipelineActivity_selectAttributes :: Lens' PipelineActivity (Maybe SelectAttributesActivity)

Used to create a new message using only the specified attributes from the original message.

pipelineActivity_channel :: Lens' PipelineActivity (Maybe ChannelActivity)

Determines the source of the messages to be processed.

pipelineActivity_addAttributes :: Lens' PipelineActivity (Maybe AddAttributesActivity)

Adds other attributes based on existing attributes in the message.

pipelineActivity_deviceRegistryEnrich :: Lens' PipelineActivity (Maybe DeviceRegistryEnrichActivity)

Adds data from the IoT device registry to your message.

pipelineActivity_removeAttributes :: Lens' PipelineActivity (Maybe RemoveAttributesActivity)

Removes attributes from a message.

pipelineActivity_lambda :: Lens' PipelineActivity (Maybe LambdaActivity)

Runs a Lambda function to modify the message.

pipelineActivity_datastore :: Lens' PipelineActivity (Maybe DatastoreActivity)

Specifies where to store the processed message data.

pipelineActivity_deviceShadowEnrich :: Lens' PipelineActivity (Maybe DeviceShadowEnrichActivity)

Adds information from the IoT Device Shadow service to a message.

pipelineActivity_filter :: Lens' PipelineActivity (Maybe FilterActivity)

Filters a message based on its attributes.

pipelineActivity_math :: Lens' PipelineActivity (Maybe MathActivity)

Computes an arithmetic expression using the message's attributes and adds it to the message.