Amazonka.IoTAnalytics.Types.Variable

Description

Documentation

data Variable

An instance of a variable to be passed to the containerAction execution. Each variable must have a name and a value given by one of stringValue, datasetContentVersionValue, or outputFileUriValue.

See: newVariable smart constructor.

Constructors

Variable'

Fields

outputFileUriValue :: Maybe OutputFileUriValue

The value of the variable as a structure that specifies an output file URI.

doubleValue :: Maybe Double

The value of the variable as a double (numeric).

stringValue :: Maybe Text

The value of the variable as a string.

datasetContentVersionValue :: Maybe DatasetContentVersionValue

The value of the variable as a structure that specifies a dataset content version.

name :: Text

The name of the variable.

newVariable

Arguments

:: Text
$sel:name:Variable'

-> Variable

Create a value of Variable with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:outputFileUriValue:Variable', variable_outputFileUriValue - The value of the variable as a structure that specifies an output file URI.

$sel:doubleValue:Variable', variable_doubleValue - The value of the variable as a double (numeric).

$sel:stringValue:Variable', variable_stringValue - The value of the variable as a string.

$sel:datasetContentVersionValue:Variable', variable_datasetContentVersionValue - The value of the variable as a structure that specifies a dataset content version.

$sel:name:Variable', variable_name - The name of the variable.

variable_outputFileUriValue :: Lens' Variable (Maybe OutputFileUriValue)

The value of the variable as a structure that specifies an output file URI.

variable_doubleValue :: Lens' Variable (Maybe Double)

The value of the variable as a double (numeric).

variable_stringValue :: Lens' Variable (Maybe Text)

The value of the variable as a string.

variable_datasetContentVersionValue :: Lens' Variable (Maybe DatasetContentVersionValue)

The value of the variable as a structure that specifies a dataset content version.

variable_name :: Lens' Variable Text

The name of the variable.