Amazonka.IoTAnalytics.UpdateDataset

Description

Updates the settings of a dataset.

Creating a Request

data UpdateDataset

See: newUpdateDataset smart constructor.

Constructors

UpdateDataset'

Fields

versioningConfiguration :: Maybe VersioningConfiguration

Optional. How many versions of dataset contents are kept. If not specified or set to null, only the latest version plus the latest succeeded version (if they are different) are kept for the time period specified by the retentionPeriod parameter. For more information, see Keeping Multiple Versions of IoT Analytics datasets in the IoT Analytics User Guide.

triggers :: Maybe [DatasetTrigger]

A list of DatasetTrigger objects. The list can be empty or can contain up to five DatasetTrigger objects.

retentionPeriod :: Maybe RetentionPeriod

How long, in days, dataset contents are kept for the dataset.

lateDataRules :: Maybe (NonEmpty LateDataRule)

A list of data rules that send notifications to CloudWatch, when data arrives late. To specify lateDataRules, the dataset must use a DeltaTimer filter.

contentDeliveryRules :: Maybe [DatasetContentDeliveryRule]

When dataset contents are created, they are delivered to destinations specified here.

datasetName :: Text

The name of the dataset to update.

actions :: NonEmpty DatasetAction

A list of DatasetAction objects.

newUpdateDataset

Arguments

:: Text
$sel:datasetName:UpdateDataset'

-> NonEmpty DatasetAction
$sel:actions:UpdateDataset'

-> UpdateDataset

Create a value of UpdateDataset with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:versioningConfiguration:UpdateDataset', updateDataset_versioningConfiguration - Optional. How many versions of dataset contents are kept. If not specified or set to null, only the latest version plus the latest succeeded version (if they are different) are kept for the time period specified by the retentionPeriod parameter. For more information, see Keeping Multiple Versions of IoT Analytics datasets in the IoT Analytics User Guide.

$sel:triggers:UpdateDataset', updateDataset_triggers - A list of DatasetTrigger objects. The list can be empty or can contain up to five DatasetTrigger objects.

$sel:retentionPeriod:UpdateDataset', updateDataset_retentionPeriod - How long, in days, dataset contents are kept for the dataset.

$sel:lateDataRules:UpdateDataset', updateDataset_lateDataRules - A list of data rules that send notifications to CloudWatch, when data arrives late. To specify lateDataRules, the dataset must use a DeltaTimer filter.

$sel:contentDeliveryRules:UpdateDataset', updateDataset_contentDeliveryRules - When dataset contents are created, they are delivered to destinations specified here.

$sel:datasetName:UpdateDataset', updateDataset_datasetName - The name of the dataset to update.

$sel:actions:UpdateDataset', updateDataset_actions - A list of DatasetAction objects.

Request Lenses

updateDataset_versioningConfiguration :: Lens' UpdateDataset (Maybe VersioningConfiguration)

Optional. How many versions of dataset contents are kept. If not specified or set to null, only the latest version plus the latest succeeded version (if they are different) are kept for the time period specified by the retentionPeriod parameter. For more information, see Keeping Multiple Versions of IoT Analytics datasets in the IoT Analytics User Guide.

updateDataset_triggers :: Lens' UpdateDataset (Maybe [DatasetTrigger])

A list of DatasetTrigger objects. The list can be empty or can contain up to five DatasetTrigger objects.

updateDataset_retentionPeriod :: Lens' UpdateDataset (Maybe RetentionPeriod)

How long, in days, dataset contents are kept for the dataset.

updateDataset_lateDataRules :: Lens' UpdateDataset (Maybe (NonEmpty LateDataRule))

A list of data rules that send notifications to CloudWatch, when data arrives late. To specify lateDataRules, the dataset must use a DeltaTimer filter.

updateDataset_contentDeliveryRules :: Lens' UpdateDataset (Maybe [DatasetContentDeliveryRule])

When dataset contents are created, they are delivered to destinations specified here.

updateDataset_datasetName :: Lens' UpdateDataset Text

The name of the dataset to update.

updateDataset_actions :: Lens' UpdateDataset (NonEmpty DatasetAction)

A list of DatasetAction objects.

Destructuring the Response

data UpdateDatasetResponse

See: newUpdateDatasetResponse smart constructor.

Constructors

UpdateDatasetResponse'

newUpdateDatasetResponse :: UpdateDatasetResponse

Create a value of UpdateDatasetResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.