Amazonka.IoTAnalytics.UpdateDatastore

Description

Used to update the settings of a data store.

Creating a Request

data UpdateDatastore

See: newUpdateDatastore smart constructor.

Constructors

UpdateDatastore'

Fields

fileFormatConfiguration :: Maybe FileFormatConfiguration

Contains the configuration information of file formats. IoT Analytics data stores support JSON and Parquet.

The default file format is JSON. You can specify only one format.

You can't change the file format after you create the data store.

retentionPeriod :: Maybe RetentionPeriod

How long, in days, message data is kept for the data store. The retention period can't be updated if the data store's Amazon S3 storage is customer-managed.

datastoreStorage :: Maybe DatastoreStorage

Where data in a data store is stored.. You can choose serviceManagedS3 storage, customerManagedS3 storage, or iotSiteWiseMultiLayerStorage storage. The default is serviceManagedS3. You can't change the choice of Amazon S3 storage after your data store is created.

datastoreName :: Text

The name of the data store to be updated.

newUpdateDatastore

Arguments

:: Text
$sel:datastoreName:UpdateDatastore'

-> UpdateDatastore

Create a value of UpdateDatastore with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:fileFormatConfiguration:UpdateDatastore', updateDatastore_fileFormatConfiguration - Contains the configuration information of file formats. IoT Analytics data stores support JSON and Parquet.

The default file format is JSON. You can specify only one format.

You can't change the file format after you create the data store.

$sel:retentionPeriod:UpdateDatastore', updateDatastore_retentionPeriod - How long, in days, message data is kept for the data store. The retention period can't be updated if the data store's Amazon S3 storage is customer-managed.

$sel:datastoreStorage:UpdateDatastore', updateDatastore_datastoreStorage - Where data in a data store is stored.. You can choose serviceManagedS3 storage, customerManagedS3 storage, or iotSiteWiseMultiLayerStorage storage. The default is serviceManagedS3. You can't change the choice of Amazon S3 storage after your data store is created.

$sel:datastoreName:UpdateDatastore', updateDatastore_datastoreName - The name of the data store to be updated.

Request Lenses

updateDatastore_fileFormatConfiguration :: Lens' UpdateDatastore (Maybe FileFormatConfiguration)

Contains the configuration information of file formats. IoT Analytics data stores support JSON and Parquet.

The default file format is JSON. You can specify only one format.

You can't change the file format after you create the data store.

updateDatastore_retentionPeriod :: Lens' UpdateDatastore (Maybe RetentionPeriod)

How long, in days, message data is kept for the data store. The retention period can't be updated if the data store's Amazon S3 storage is customer-managed.

updateDatastore_datastoreStorage :: Lens' UpdateDatastore (Maybe DatastoreStorage)

Where data in a data store is stored.. You can choose serviceManagedS3 storage, customerManagedS3 storage, or iotSiteWiseMultiLayerStorage storage. The default is serviceManagedS3. You can't change the choice of Amazon S3 storage after your data store is created.

updateDatastore_datastoreName :: Lens' UpdateDatastore Text

The name of the data store to be updated.

Destructuring the Response

data UpdateDatastoreResponse

See: newUpdateDatastoreResponse smart constructor.

Constructors

UpdateDatastoreResponse'

newUpdateDatastoreResponse :: UpdateDatastoreResponse

Create a value of UpdateDatastoreResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.