Amazonka.IoTData.DeleteThingShadow

Description

Deletes the shadow for the specified thing.

Requires permission to access the DeleteThingShadow action.

For more information, see DeleteThingShadow in the IoT Developer Guide.

Creating a Request

data DeleteThingShadow

The input for the DeleteThingShadow operation.

See: newDeleteThingShadow smart constructor.

Constructors

DeleteThingShadow'

Fields

shadowName :: Maybe Text

The name of the shadow.

thingName :: Text

The name of the thing.

newDeleteThingShadow

Arguments

:: Text
$sel:thingName:DeleteThingShadow'

-> DeleteThingShadow

Create a value of DeleteThingShadow with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:shadowName:DeleteThingShadow', deleteThingShadow_shadowName - The name of the shadow.

$sel:thingName:DeleteThingShadow', deleteThingShadow_thingName - The name of the thing.

Request Lenses

deleteThingShadow_shadowName :: Lens' DeleteThingShadow (Maybe Text)

The name of the shadow.

deleteThingShadow_thingName :: Lens' DeleteThingShadow Text

The name of the thing.

Destructuring the Response

data DeleteThingShadowResponse

The output from the DeleteThingShadow operation.

See: newDeleteThingShadowResponse smart constructor.

Constructors

DeleteThingShadowResponse'

Fields

httpStatus :: Int

The response's http status code.

payload :: ByteString

The state information, in JSON format.

newDeleteThingShadowResponse

Arguments

:: Int
$sel:httpStatus:DeleteThingShadowResponse'

-> ByteString
$sel:payload:DeleteThingShadowResponse'

-> DeleteThingShadowResponse

Create a value of DeleteThingShadowResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:httpStatus:DeleteThingShadowResponse', deleteThingShadowResponse_httpStatus - The response's http status code.

$sel:payload:DeleteThingShadowResponse', deleteThingShadowResponse_payload - The state information, in JSON format.

Response Lenses

deleteThingShadowResponse_httpStatus :: Lens' DeleteThingShadowResponse Int

The response's http status code.

deleteThingShadowResponse_payload :: Lens' DeleteThingShadowResponse ByteString

The state information, in JSON format.