Amazonka.IoTDeviceAdvisor

Description

Derived from API version 2020-09-18 of the AWS service descriptions, licensed under Apache 2.0.

AWS IoT Core Device Advisor is a cloud-based, fully managed test capability for validating IoT devices during device software development. Device Advisor provides pre-built tests that you can use to validate IoT devices for reliable and secure connectivity with AWS IoT Core before deploying devices to production. By using Device Advisor, you can confirm that your devices can connect to AWS IoT Core, follow security best practices and, if applicable, receive software updates from IoT Device Management. You can also download signed qualification reports to submit to the AWS Partner Network to get your device qualified for the AWS Partner Device Catalog without the need to send your device in and wait for it to be tested.

Service Configuration

defaultService :: Service

API version 2020-09-18 of the Amazon IoT Core Device Advisor SDK configuration.

Errors

Error matchers are designed for use with the functions provided by Control.Exception.Lens. This allows catching (and rethrowing) service specific errors returned by IoTDeviceAdvisor.

ValidationException

_ValidationException :: AsError a => Getting (First ServiceError) a ServiceError

Sends invalid request exception.

ConflictException

_ConflictException :: AsError a => Getting (First ServiceError) a ServiceError

Sends Conflict Exception.

InternalServerException

_InternalServerException :: AsError a => Getting (First ServiceError) a ServiceError

Sends Internal Failure Exception.

ResourceNotFoundException

_ResourceNotFoundException :: AsError a => Getting (First ServiceError) a ServiceError

Sends Resource Not Found Exception.

Waiters

Waiters poll by repeatedly sending a request until some remote success condition configured by the Wait specification is fulfilled. The Wait specification determines how many attempts should be made, in addition to delay and retry strategies.

Operations

Some AWS operations return results that are incomplete and require subsequent requests in order to obtain the entire result set. The process of sending subsequent requests to continue where a previous request left off is called pagination. For example, the ListObjects operation of Amazon S3 returns up to 1000 objects at a time, and you must send subsequent requests with the appropriate Marker in order to retrieve the next page of results.

Operations that have an AWSPager instance can transparently perform subsequent requests, correctly setting Markers and other request facets to iterate through the entire result set of a truncated API operation. Operations which support this have an additional note in the documentation.

Many operations have the ability to filter results on the server side. See the individual operation parameters for details.

GetSuiteRunReport

data GetSuiteRunReport

See: newGetSuiteRunReport smart constructor.

Constructors

GetSuiteRunReport' Text Text

newGetSuiteRunReport

Arguments

:: Text
$sel:suiteDefinitionId:GetSuiteRunReport'

-> Text
$sel:suiteRunId:GetSuiteRunReport'

-> GetSuiteRunReport

Create a value of GetSuiteRunReport with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:suiteDefinitionId:GetSuiteRunReport', getSuiteRunReport_suiteDefinitionId - Suite definition Id of the test suite.

$sel:suiteRunId:GetSuiteRunReport', getSuiteRunReport_suiteRunId - Suite run Id of the test suite run.

data GetSuiteRunReportResponse

See: newGetSuiteRunReportResponse smart constructor.

Constructors

GetSuiteRunReportResponse' (Maybe Text) Int

newGetSuiteRunReportResponse

Arguments

:: Int
$sel:httpStatus:GetSuiteRunReportResponse'

-> GetSuiteRunReportResponse

Create a value of GetSuiteRunReportResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:qualificationReportDownloadUrl:GetSuiteRunReportResponse', getSuiteRunReportResponse_qualificationReportDownloadUrl - Download URL of the qualification report.

$sel:httpStatus:GetSuiteRunReportResponse', getSuiteRunReportResponse_httpStatus - The response's http status code.

StartSuiteRun

data StartSuiteRun

See: newStartSuiteRun smart constructor.

Constructors

StartSuiteRun' (Maybe SuiteRunConfiguration) (Maybe Text) (Maybe (HashMap Text Text)) Text

newStartSuiteRun

Arguments

:: Text
$sel:suiteDefinitionId:StartSuiteRun'

-> StartSuiteRun

Create a value of StartSuiteRun with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:suiteRunConfiguration:StartSuiteRun', startSuiteRun_suiteRunConfiguration - Suite run configuration.

$sel:suiteDefinitionVersion:StartSuiteRun', startSuiteRun_suiteDefinitionVersion - Suite definition version of the test suite.

$sel:tags:StartSuiteRun', startSuiteRun_tags - The tags to be attached to the suite run.

$sel:suiteDefinitionId:StartSuiteRun', startSuiteRun_suiteDefinitionId - Suite definition Id of the test suite.

data StartSuiteRunResponse

See: newStartSuiteRunResponse smart constructor.

Constructors

StartSuiteRunResponse' (Maybe POSIX) (Maybe Text) (Maybe Text) Int

newStartSuiteRunResponse

Arguments

:: Int
$sel:httpStatus:StartSuiteRunResponse'

-> StartSuiteRunResponse

Create a value of StartSuiteRunResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:createdAt:StartSuiteRunResponse', startSuiteRunResponse_createdAt - Date (in Unix epoch time) when the suite run was created.

$sel:suiteRunArn:StartSuiteRunResponse', startSuiteRunResponse_suiteRunArn - Amazon resource name of the started suite run.

$sel:suiteRunId:StartSuiteRunResponse', startSuiteRunResponse_suiteRunId - Suite Run Id of the started suite run.

$sel:httpStatus:StartSuiteRunResponse', startSuiteRunResponse_httpStatus - The response's http status code.

ListSuiteDefinitions

data ListSuiteDefinitions

See: newListSuiteDefinitions smart constructor.

Constructors

ListSuiteDefinitions' (Maybe Text) (Maybe Natural)

newListSuiteDefinitions :: ListSuiteDefinitions

Create a value of ListSuiteDefinitions with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:nextToken:ListSuiteDefinitions', listSuiteDefinitions_nextToken - A token used to get the next set of results.

$sel:maxResults:ListSuiteDefinitions', listSuiteDefinitions_maxResults - The maximum number of results to return at once.

data ListSuiteDefinitionsResponse

See: newListSuiteDefinitionsResponse smart constructor.

Constructors

ListSuiteDefinitionsResponse' (Maybe [SuiteDefinitionInformation]) (Maybe Text) Int

newListSuiteDefinitionsResponse

Arguments

:: Int
$sel:httpStatus:ListSuiteDefinitionsResponse'

-> ListSuiteDefinitionsResponse

Create a value of ListSuiteDefinitionsResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:suiteDefinitionInformationList:ListSuiteDefinitionsResponse', listSuiteDefinitionsResponse_suiteDefinitionInformationList - An array of objects that provide summaries of information about the suite definitions in the list.

$sel:nextToken:ListSuiteDefinitions', listSuiteDefinitionsResponse_nextToken - A token used to get the next set of results.

$sel:httpStatus:ListSuiteDefinitionsResponse', listSuiteDefinitionsResponse_httpStatus - The response's http status code.

ListTagsForResource

data ListTagsForResource

See: newListTagsForResource smart constructor.

Constructors

ListTagsForResource' Text

newListTagsForResource

Arguments

:: Text
$sel:resourceArn:ListTagsForResource'

-> ListTagsForResource

Create a value of ListTagsForResource with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:resourceArn:ListTagsForResource', listTagsForResource_resourceArn - The ARN of the IoT Device Advisor resource.

data ListTagsForResourceResponse

See: newListTagsForResourceResponse smart constructor.

Constructors

ListTagsForResourceResponse' (Maybe (HashMap Text Text)) Int

newListTagsForResourceResponse

Arguments

:: Int
$sel:httpStatus:ListTagsForResourceResponse'

-> ListTagsForResourceResponse

Create a value of ListTagsForResourceResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:tags:ListTagsForResourceResponse', listTagsForResourceResponse_tags - The tags attached to the IoT Device Advisor resource.

$sel:httpStatus:ListTagsForResourceResponse', listTagsForResourceResponse_httpStatus - The response's http status code.

DeleteSuiteDefinition

data DeleteSuiteDefinition

See: newDeleteSuiteDefinition smart constructor.

Constructors

DeleteSuiteDefinition' Text

newDeleteSuiteDefinition

Arguments

:: Text
$sel:suiteDefinitionId:DeleteSuiteDefinition'

-> DeleteSuiteDefinition

Create a value of DeleteSuiteDefinition with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:suiteDefinitionId:DeleteSuiteDefinition', deleteSuiteDefinition_suiteDefinitionId - Suite definition Id of the test suite to be deleted.

data DeleteSuiteDefinitionResponse

See: newDeleteSuiteDefinitionResponse smart constructor.

Constructors

DeleteSuiteDefinitionResponse' Int

newDeleteSuiteDefinitionResponse

Arguments

:: Int
$sel:httpStatus:DeleteSuiteDefinitionResponse'

-> DeleteSuiteDefinitionResponse

Create a value of DeleteSuiteDefinitionResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:httpStatus:DeleteSuiteDefinitionResponse', deleteSuiteDefinitionResponse_httpStatus - The response's http status code.

UpdateSuiteDefinition

data UpdateSuiteDefinition

See: newUpdateSuiteDefinition smart constructor.

Constructors

UpdateSuiteDefinition' (Maybe SuiteDefinitionConfiguration) Text

newUpdateSuiteDefinition

Arguments

:: Text
$sel:suiteDefinitionId:UpdateSuiteDefinition'

-> UpdateSuiteDefinition

Create a value of UpdateSuiteDefinition with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:suiteDefinitionConfiguration:UpdateSuiteDefinition', updateSuiteDefinition_suiteDefinitionConfiguration - Updates a Device Advisor test suite with suite definition configuration.

$sel:suiteDefinitionId:UpdateSuiteDefinition', updateSuiteDefinition_suiteDefinitionId - Suite definition Id of the test suite to be updated.

data UpdateSuiteDefinitionResponse

See: newUpdateSuiteDefinitionResponse smart constructor.

Constructors

UpdateSuiteDefinitionResponse' (Maybe POSIX) (Maybe POSIX) (Maybe Text) (Maybe Text) (Maybe Text) (Maybe Text) Int

newUpdateSuiteDefinitionResponse

Arguments

:: Int
$sel:httpStatus:UpdateSuiteDefinitionResponse'

-> UpdateSuiteDefinitionResponse

Create a value of UpdateSuiteDefinitionResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:lastUpdatedAt:UpdateSuiteDefinitionResponse', updateSuiteDefinitionResponse_lastUpdatedAt - Timestamp of when the test suite was updated.

$sel:createdAt:UpdateSuiteDefinitionResponse', updateSuiteDefinitionResponse_createdAt - Timestamp of when the test suite was created.

$sel:suiteDefinitionArn:UpdateSuiteDefinitionResponse', updateSuiteDefinitionResponse_suiteDefinitionArn - Amazon Resource name of the updated test suite.

$sel:suiteDefinitionId:UpdateSuiteDefinition', updateSuiteDefinitionResponse_suiteDefinitionId - Suite definition Id of the updated test suite.

$sel:suiteDefinitionVersion:UpdateSuiteDefinitionResponse', updateSuiteDefinitionResponse_suiteDefinitionVersion - Suite definition version of the updated test suite.

$sel:suiteDefinitionName:UpdateSuiteDefinitionResponse', updateSuiteDefinitionResponse_suiteDefinitionName - Suite definition name of the updated test suite.

$sel:httpStatus:UpdateSuiteDefinitionResponse', updateSuiteDefinitionResponse_httpStatus - The response's http status code.

CreateSuiteDefinition

data CreateSuiteDefinition

See: newCreateSuiteDefinition smart constructor.

Constructors

CreateSuiteDefinition' (Maybe SuiteDefinitionConfiguration) (Maybe (HashMap Text Text))

newCreateSuiteDefinition :: CreateSuiteDefinition

Create a value of CreateSuiteDefinition with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:suiteDefinitionConfiguration:CreateSuiteDefinition', createSuiteDefinition_suiteDefinitionConfiguration - Creates a Device Advisor test suite with suite definition configuration.

$sel:tags:CreateSuiteDefinition', createSuiteDefinition_tags - The tags to be attached to the suite definition.

data CreateSuiteDefinitionResponse

See: newCreateSuiteDefinitionResponse smart constructor.

Constructors

CreateSuiteDefinitionResponse' (Maybe POSIX) (Maybe Text) (Maybe Text) (Maybe Text) Int

newCreateSuiteDefinitionResponse

Arguments

:: Int
$sel:httpStatus:CreateSuiteDefinitionResponse'

-> CreateSuiteDefinitionResponse

Create a value of CreateSuiteDefinitionResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:createdAt:CreateSuiteDefinitionResponse', createSuiteDefinitionResponse_createdAt - Creates a Device Advisor test suite with TimeStamp of when it was created.

$sel:suiteDefinitionArn:CreateSuiteDefinitionResponse', createSuiteDefinitionResponse_suiteDefinitionArn - Creates a Device Advisor test suite with Amazon Resource name.

$sel:suiteDefinitionId:CreateSuiteDefinitionResponse', createSuiteDefinitionResponse_suiteDefinitionId - Creates a Device Advisor test suite with suite UUID.

$sel:suiteDefinitionName:CreateSuiteDefinitionResponse', createSuiteDefinitionResponse_suiteDefinitionName - Creates a Device Advisor test suite with suite definition name.

$sel:httpStatus:CreateSuiteDefinitionResponse', createSuiteDefinitionResponse_httpStatus - The response's http status code.

StopSuiteRun

data StopSuiteRun

See: newStopSuiteRun smart constructor.

Constructors

StopSuiteRun' Text Text

newStopSuiteRun

Arguments

:: Text
$sel:suiteDefinitionId:StopSuiteRun'

-> Text
$sel:suiteRunId:StopSuiteRun'

-> StopSuiteRun

Create a value of StopSuiteRun with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:suiteDefinitionId:StopSuiteRun', stopSuiteRun_suiteDefinitionId - Suite definition Id of the test suite run to be stopped.

$sel:suiteRunId:StopSuiteRun', stopSuiteRun_suiteRunId - Suite run Id of the test suite run to be stopped.

data StopSuiteRunResponse

See: newStopSuiteRunResponse smart constructor.

Constructors

StopSuiteRunResponse' Int

newStopSuiteRunResponse

Arguments

:: Int
$sel:httpStatus:StopSuiteRunResponse'

-> StopSuiteRunResponse

Create a value of StopSuiteRunResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:httpStatus:StopSuiteRunResponse', stopSuiteRunResponse_httpStatus - The response's http status code.

GetSuiteDefinition

data GetSuiteDefinition

See: newGetSuiteDefinition smart constructor.

Constructors

GetSuiteDefinition' (Maybe Text) Text

newGetSuiteDefinition

Arguments

:: Text
$sel:suiteDefinitionId:GetSuiteDefinition'

-> GetSuiteDefinition

Create a value of GetSuiteDefinition with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:suiteDefinitionVersion:GetSuiteDefinition', getSuiteDefinition_suiteDefinitionVersion - Suite definition version of the test suite to get.

$sel:suiteDefinitionId:GetSuiteDefinition', getSuiteDefinition_suiteDefinitionId - Suite definition Id of the test suite to get.

data GetSuiteDefinitionResponse

See: newGetSuiteDefinitionResponse smart constructor.

Constructors

GetSuiteDefinitionResponse' (Maybe POSIX) (Maybe SuiteDefinitionConfiguration) (Maybe Text) (Maybe POSIX) (Maybe Text) (Maybe Text) (Maybe Text) (Maybe (HashMap Text Text)) Int

newGetSuiteDefinitionResponse

Arguments

:: Int
$sel:httpStatus:GetSuiteDefinitionResponse'

-> GetSuiteDefinitionResponse

Create a value of GetSuiteDefinitionResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:createdAt:GetSuiteDefinitionResponse', getSuiteDefinitionResponse_createdAt - Date (in Unix epoch time) when the suite definition was created.

$sel:suiteDefinitionConfiguration:GetSuiteDefinitionResponse', getSuiteDefinitionResponse_suiteDefinitionConfiguration - Suite configuration of the suite definition.

$sel:suiteDefinitionArn:GetSuiteDefinitionResponse', getSuiteDefinitionResponse_suiteDefinitionArn - The ARN of the suite definition.

$sel:lastModifiedAt:GetSuiteDefinitionResponse', getSuiteDefinitionResponse_lastModifiedAt - Date (in Unix epoch time) when the suite definition was last modified.

$sel:suiteDefinitionId:GetSuiteDefinition', getSuiteDefinitionResponse_suiteDefinitionId - Suite definition Id of the suite definition.

$sel:suiteDefinitionVersion:GetSuiteDefinition', getSuiteDefinitionResponse_suiteDefinitionVersion - Suite definition version of the suite definition.

$sel:latestVersion:GetSuiteDefinitionResponse', getSuiteDefinitionResponse_latestVersion - Latest suite definition version of the suite definition.

$sel:tags:GetSuiteDefinitionResponse', getSuiteDefinitionResponse_tags - Tags attached to the suite definition.

$sel:httpStatus:GetSuiteDefinitionResponse', getSuiteDefinitionResponse_httpStatus - The response's http status code.

TagResource

data TagResource

See: newTagResource smart constructor.

Constructors

TagResource' Text (HashMap Text Text)

newTagResource

Arguments

:: Text
$sel:resourceArn:TagResource'

-> TagResource

Create a value of TagResource with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:resourceArn:TagResource', tagResource_resourceArn - The resource ARN of an IoT Device Advisor resource.

$sel:tags:TagResource', tagResource_tags - The tags to be attached to the IoT Device Advisor resource.

data TagResourceResponse

See: newTagResourceResponse smart constructor.

Constructors

TagResourceResponse' Int

newTagResourceResponse

Arguments

:: Int
$sel:httpStatus:TagResourceResponse'

-> TagResourceResponse

Create a value of TagResourceResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:httpStatus:TagResourceResponse', tagResourceResponse_httpStatus - The response's http status code.

UntagResource

data UntagResource

See: newUntagResource smart constructor.

Constructors

UntagResource' Text [Text]

newUntagResource

Arguments

:: Text
$sel:resourceArn:UntagResource'

-> UntagResource

Create a value of UntagResource with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:resourceArn:UntagResource', untagResource_resourceArn - The resource ARN of an IoT Device Advisor resource.

$sel:tagKeys:UntagResource', untagResource_tagKeys - List of tag keys to remove from the IoT Device Advisor resource.

data UntagResourceResponse

See: newUntagResourceResponse smart constructor.

Constructors

UntagResourceResponse' Int

newUntagResourceResponse

Arguments

:: Int
$sel:httpStatus:UntagResourceResponse'

-> UntagResourceResponse

Create a value of UntagResourceResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:httpStatus:UntagResourceResponse', untagResourceResponse_httpStatus - The response's http status code.

ListSuiteRuns

data ListSuiteRuns

See: newListSuiteRuns smart constructor.

Constructors

ListSuiteRuns' (Maybe Text) (Maybe Text) (Maybe Text) (Maybe Natural)

newListSuiteRuns :: ListSuiteRuns

Create a value of ListSuiteRuns with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:suiteDefinitionId:ListSuiteRuns', listSuiteRuns_suiteDefinitionId - Lists the test suite runs of the specified test suite based on suite definition Id.

$sel:suiteDefinitionVersion:ListSuiteRuns', listSuiteRuns_suiteDefinitionVersion - Must be passed along with suiteDefinitionId. Lists the test suite runs of the specified test suite based on suite definition version.

$sel:nextToken:ListSuiteRuns', listSuiteRuns_nextToken - A token to retrieve the next set of results.

$sel:maxResults:ListSuiteRuns', listSuiteRuns_maxResults - The maximum number of results to return at once.

data ListSuiteRunsResponse

See: newListSuiteRunsResponse smart constructor.

Constructors

ListSuiteRunsResponse' (Maybe Text) (Maybe [SuiteRunInformation]) Int

newListSuiteRunsResponse

Arguments

:: Int
$sel:httpStatus:ListSuiteRunsResponse'

-> ListSuiteRunsResponse

Create a value of ListSuiteRunsResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:nextToken:ListSuiteRuns', listSuiteRunsResponse_nextToken - A token to retrieve the next set of results.

$sel:suiteRunsList:ListSuiteRunsResponse', listSuiteRunsResponse_suiteRunsList - An array of objects that provide summaries of information about the suite runs in the list.

$sel:httpStatus:ListSuiteRunsResponse', listSuiteRunsResponse_httpStatus - The response's http status code.

GetSuiteRun

data GetSuiteRun

See: newGetSuiteRun smart constructor.

Constructors

GetSuiteRun' Text Text

newGetSuiteRun

Arguments

:: Text
$sel:suiteDefinitionId:GetSuiteRun'

-> Text
$sel:suiteRunId:GetSuiteRun'

-> GetSuiteRun

Create a value of GetSuiteRun with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:suiteDefinitionId:GetSuiteRun', getSuiteRun_suiteDefinitionId - Suite definition Id for the test suite run.

$sel:suiteRunId:GetSuiteRun', getSuiteRun_suiteRunId - Suite run Id for the test suite run.

data GetSuiteRunResponse

See: newGetSuiteRunResponse smart constructor.

Constructors

GetSuiteRunResponse' (Maybe SuiteRunStatus) (Maybe SuiteRunConfiguration) (Maybe POSIX) (Maybe Text) (Maybe Text) (Maybe Text) (Maybe TestResult) (Maybe POSIX) (Maybe Text) (Maybe Text) (Maybe (HashMap Text Text)) Int

newGetSuiteRunResponse

Arguments

:: Int
$sel:httpStatus:GetSuiteRunResponse'

-> GetSuiteRunResponse

Create a value of GetSuiteRunResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:status:GetSuiteRunResponse', getSuiteRunResponse_status - Status for the test suite run.

$sel:suiteRunConfiguration:GetSuiteRunResponse', getSuiteRunResponse_suiteRunConfiguration - Suite run configuration for the test suite run.

$sel:startTime:GetSuiteRunResponse', getSuiteRunResponse_startTime - Date (in Unix epoch time) when the test suite run was started.

$sel:suiteDefinitionId:GetSuiteRun', getSuiteRunResponse_suiteDefinitionId - Suite definition Id for the test suite run.

$sel:suiteDefinitionVersion:GetSuiteRunResponse', getSuiteRunResponse_suiteDefinitionVersion - Suite definition version for the test suite run.

$sel:errorReason:GetSuiteRunResponse', getSuiteRunResponse_errorReason - Error reason for any test suite run failure.

$sel:testResult:GetSuiteRunResponse', getSuiteRunResponse_testResult - Test results for the test suite run.

$sel:endTime:GetSuiteRunResponse', getSuiteRunResponse_endTime - Date (in Unix epoch time) when the test suite run ended.

$sel:suiteRunArn:GetSuiteRunResponse', getSuiteRunResponse_suiteRunArn - The ARN of the suite run.

$sel:suiteRunId:GetSuiteRun', getSuiteRunResponse_suiteRunId - Suite run Id for the test suite run.

$sel:tags:GetSuiteRunResponse', getSuiteRunResponse_tags - The tags attached to the suite run.

$sel:httpStatus:GetSuiteRunResponse', getSuiteRunResponse_httpStatus - The response's http status code.

Types

Status

newtype Status

Constructors

Status'

Fields

fromStatus :: Text

Bundled Patterns

pattern Status_CANCELED :: Status

pattern Status_ERROR :: Status

pattern Status_FAIL :: Status

pattern Status_PASS :: Status

pattern Status_PASS_WITH_WARNINGS :: Status

pattern Status_PENDING :: Status

pattern Status_RUNNING :: Status

pattern Status_STOPPED :: Status

pattern Status_STOPPING :: Status

SuiteRunStatus

newtype SuiteRunStatus

Constructors

SuiteRunStatus'

Fields

fromSuiteRunStatus :: Text

Bundled Patterns

pattern SuiteRunStatus_CANCELED :: SuiteRunStatus

pattern SuiteRunStatus_ERROR :: SuiteRunStatus

pattern SuiteRunStatus_FAIL :: SuiteRunStatus

pattern SuiteRunStatus_PASS :: SuiteRunStatus

pattern SuiteRunStatus_PASS_WITH_WARNINGS :: SuiteRunStatus

pattern SuiteRunStatus_PENDING :: SuiteRunStatus

pattern SuiteRunStatus_RUNNING :: SuiteRunStatus

pattern SuiteRunStatus_STOPPED :: SuiteRunStatus

pattern SuiteRunStatus_STOPPING :: SuiteRunStatus

DeviceUnderTest

data DeviceUnderTest

Lists all the devices under test

See: newDeviceUnderTest smart constructor.

Constructors

DeviceUnderTest' (Maybe Text) (Maybe Text)

newDeviceUnderTest :: DeviceUnderTest

Create a value of DeviceUnderTest with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:certificateArn:DeviceUnderTest', deviceUnderTest_certificateArn - Lists devices certificate arn

$sel:thingArn:DeviceUnderTest', deviceUnderTest_thingArn - Lists devices thing arn

GroupResult

data GroupResult

Show Group Result.

See: newGroupResult smart constructor.

Constructors

GroupResult' (Maybe [TestCaseRun]) (Maybe Text) (Maybe Text)

newGroupResult :: GroupResult

Create a value of GroupResult with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:tests:GroupResult', groupResult_tests - Tests under Group Result.

$sel:groupId:GroupResult', groupResult_groupId - Group result Id.

$sel:groupName:GroupResult', groupResult_groupName - Group Result Name.

SuiteDefinitionConfiguration

data SuiteDefinitionConfiguration

Gets Suite Definition Configuration.

See: newSuiteDefinitionConfiguration smart constructor.

Constructors

SuiteDefinitionConfiguration' (Maybe Text) (Maybe Bool) (Maybe Text) (Maybe [DeviceUnderTest]) (Maybe Text)

newSuiteDefinitionConfiguration :: SuiteDefinitionConfiguration

Create a value of SuiteDefinitionConfiguration with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:suiteDefinitionName:SuiteDefinitionConfiguration', suiteDefinitionConfiguration_suiteDefinitionName - Gets Suite Definition Configuration name.

$sel:intendedForQualification:SuiteDefinitionConfiguration', suiteDefinitionConfiguration_intendedForQualification - Gets the tests intended for qualification in a suite.

$sel:devicePermissionRoleArn:SuiteDefinitionConfiguration', suiteDefinitionConfiguration_devicePermissionRoleArn - Gets device permission arn.

$sel:devices:SuiteDefinitionConfiguration', suiteDefinitionConfiguration_devices - Gets the devices configured.

$sel:rootGroup:SuiteDefinitionConfiguration', suiteDefinitionConfiguration_rootGroup - Gets test suite root group.

SuiteDefinitionInformation

data SuiteDefinitionInformation

Information about the suite definition.

See: newSuiteDefinitionInformation smart constructor.

Constructors

SuiteDefinitionInformation' (Maybe POSIX) (Maybe [DeviceUnderTest]) (Maybe Text) (Maybe Text) (Maybe Bool)

newSuiteDefinitionInformation :: SuiteDefinitionInformation

Create a value of SuiteDefinitionInformation with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:createdAt:SuiteDefinitionInformation', suiteDefinitionInformation_createdAt - Date (in Unix epoch time) when the test suite was created.

$sel:defaultDevices:SuiteDefinitionInformation', suiteDefinitionInformation_defaultDevices - Specifies the devices under test for the test suite.

$sel:suiteDefinitionId:SuiteDefinitionInformation', suiteDefinitionInformation_suiteDefinitionId - Suite definition Id of the test suite.

$sel:suiteDefinitionName:SuiteDefinitionInformation', suiteDefinitionInformation_suiteDefinitionName - Suite name of the test suite.

$sel:intendedForQualification:SuiteDefinitionInformation', suiteDefinitionInformation_intendedForQualification - Specifies if the test suite is intended for qualification.

SuiteRunConfiguration

data SuiteRunConfiguration

Gets suite run configuration.

See: newSuiteRunConfiguration smart constructor.

Constructors

SuiteRunConfiguration' (Maybe DeviceUnderTest) (Maybe [Text])

newSuiteRunConfiguration :: SuiteRunConfiguration

Create a value of SuiteRunConfiguration with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:primaryDevice:SuiteRunConfiguration', suiteRunConfiguration_primaryDevice - Gets the primary device for suite run.

$sel:selectedTestList:SuiteRunConfiguration', suiteRunConfiguration_selectedTestList - Gets test case list.

SuiteRunInformation

data SuiteRunInformation

Information about the suite run.

See: newSuiteRunInformation smart constructor.

Constructors

SuiteRunInformation' (Maybe SuiteRunStatus) (Maybe POSIX) (Maybe Natural) (Maybe Text) (Maybe Text) (Maybe POSIX) (Maybe Text) (Maybe POSIX) (Maybe Text) (Maybe Natural)

newSuiteRunInformation :: SuiteRunInformation

Create a value of SuiteRunInformation with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:status:SuiteRunInformation', suiteRunInformation_status - Status of the suite run.

$sel:createdAt:SuiteRunInformation', suiteRunInformation_createdAt - Date (in Unix epoch time) when the suite run was created.

$sel:passed:SuiteRunInformation', suiteRunInformation_passed - Number of test cases that passed in the suite run.

$sel:suiteDefinitionId:SuiteRunInformation', suiteRunInformation_suiteDefinitionId - Suite definition Id of the suite run.

$sel:suiteDefinitionVersion:SuiteRunInformation', suiteRunInformation_suiteDefinitionVersion - Suite definition version of the suite run.

$sel:startedAt:SuiteRunInformation', suiteRunInformation_startedAt - Date (in Unix epoch time) when the suite run was started.

$sel:suiteDefinitionName:SuiteRunInformation', suiteRunInformation_suiteDefinitionName - Suite definition name of the suite run.

$sel:endAt:SuiteRunInformation', suiteRunInformation_endAt - Date (in Unix epoch time) when the suite run ended.

$sel:suiteRunId:SuiteRunInformation', suiteRunInformation_suiteRunId - Suite run Id of the suite run.

$sel:failed:SuiteRunInformation', suiteRunInformation_failed - Number of test cases that failed in the suite run.

TestCaseRun

data TestCaseRun

Provides test case run.

See: newTestCaseRun smart constructor.

Constructors

TestCaseRun' (Maybe Status) (Maybe Text) (Maybe POSIX) (Maybe Text) (Maybe Text) (Maybe POSIX) (Maybe Text) (Maybe Text) (Maybe Text)

newTestCaseRun :: TestCaseRun

Create a value of TestCaseRun with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:status:TestCaseRun', testCaseRun_status - Provides test case run status.

$sel:logUrl:TestCaseRun', testCaseRun_logUrl - Provides test case run log Url.

$sel:startTime:TestCaseRun', testCaseRun_startTime - Provides test case run start time.

$sel:testCaseRunId:TestCaseRun', testCaseRun_testCaseRunId - Provides test case run Id.

$sel:warnings:TestCaseRun', testCaseRun_warnings - Provides test case run warnings.

$sel:endTime:TestCaseRun', testCaseRun_endTime - Provides test case run end time.

$sel:testCaseDefinitionId:TestCaseRun', testCaseRun_testCaseDefinitionId - Provides test case run definition Id.

$sel:failure:TestCaseRun', testCaseRun_failure - Provides test case run failure result.

$sel:testCaseDefinitionName:TestCaseRun', testCaseRun_testCaseDefinitionName - Provides test case run definition Name.

TestResult

data TestResult

Show each group result.

See: newTestResult smart constructor.

Constructors

TestResult' (Maybe [GroupResult])

newTestResult :: TestResult

Create a value of TestResult with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:groups:TestResult', testResult_groups - Show each group of test results.