Amazonka.IoTDeviceAdvisor.DeleteSuiteDefinition

Description

Deletes a Device Advisor test suite.

Creating a Request

data DeleteSuiteDefinition

See: newDeleteSuiteDefinition smart constructor.

Constructors

DeleteSuiteDefinition'

Fields

suiteDefinitionId :: Text

Suite definition Id of the test suite to be deleted.

newDeleteSuiteDefinition

Arguments

:: Text
$sel:suiteDefinitionId:DeleteSuiteDefinition'

-> DeleteSuiteDefinition

Create a value of DeleteSuiteDefinition with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:suiteDefinitionId:DeleteSuiteDefinition', deleteSuiteDefinition_suiteDefinitionId - Suite definition Id of the test suite to be deleted.

Request Lenses

deleteSuiteDefinition_suiteDefinitionId :: Lens' DeleteSuiteDefinition Text

Suite definition Id of the test suite to be deleted.

Destructuring the Response

data DeleteSuiteDefinitionResponse

See: newDeleteSuiteDefinitionResponse smart constructor.

Constructors

DeleteSuiteDefinitionResponse'

Fields

httpStatus :: Int

The response's http status code.

newDeleteSuiteDefinitionResponse

Arguments

:: Int
$sel:httpStatus:DeleteSuiteDefinitionResponse'

-> DeleteSuiteDefinitionResponse

Create a value of DeleteSuiteDefinitionResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:httpStatus:DeleteSuiteDefinitionResponse', deleteSuiteDefinitionResponse_httpStatus - The response's http status code.

Response Lenses

deleteSuiteDefinitionResponse_httpStatus :: Lens' DeleteSuiteDefinitionResponse Int

The response's http status code.