Amazonka.IoTDeviceAdvisor.GetSuiteRun

Description

Gets information about a Device Advisor test suite run.

Creating a Request

data GetSuiteRun

See: newGetSuiteRun smart constructor.

Constructors

GetSuiteRun'

Fields

suiteDefinitionId :: Text

Suite definition Id for the test suite run.

suiteRunId :: Text

Suite run Id for the test suite run.

newGetSuiteRun

Arguments

:: Text
$sel:suiteDefinitionId:GetSuiteRun'

-> Text
$sel:suiteRunId:GetSuiteRun'

-> GetSuiteRun

Create a value of GetSuiteRun with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:suiteDefinitionId:GetSuiteRun', getSuiteRun_suiteDefinitionId - Suite definition Id for the test suite run.

$sel:suiteRunId:GetSuiteRun', getSuiteRun_suiteRunId - Suite run Id for the test suite run.

Request Lenses

getSuiteRun_suiteDefinitionId :: Lens' GetSuiteRun Text

Suite definition Id for the test suite run.

getSuiteRun_suiteRunId :: Lens' GetSuiteRun Text

Suite run Id for the test suite run.

Destructuring the Response

data GetSuiteRunResponse

See: newGetSuiteRunResponse smart constructor.

Constructors

GetSuiteRunResponse'

Fields

status :: Maybe SuiteRunStatus

Status for the test suite run.

suiteRunConfiguration :: Maybe SuiteRunConfiguration

Suite run configuration for the test suite run.

startTime :: Maybe POSIX

Date (in Unix epoch time) when the test suite run was started.

suiteDefinitionId :: Maybe Text

Suite definition Id for the test suite run.

suiteDefinitionVersion :: Maybe Text

Suite definition version for the test suite run.

errorReason :: Maybe Text

Error reason for any test suite run failure.

testResult :: Maybe TestResult

Test results for the test suite run.

endTime :: Maybe POSIX

Date (in Unix epoch time) when the test suite run ended.

suiteRunArn :: Maybe Text

The ARN of the suite run.

suiteRunId :: Maybe Text

Suite run Id for the test suite run.

tags :: Maybe (HashMap Text Text)

The tags attached to the suite run.

httpStatus :: Int

The response's http status code.

newGetSuiteRunResponse

Arguments

:: Int
$sel:httpStatus:GetSuiteRunResponse'

-> GetSuiteRunResponse

Create a value of GetSuiteRunResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:status:GetSuiteRunResponse', getSuiteRunResponse_status - Status for the test suite run.

$sel:suiteRunConfiguration:GetSuiteRunResponse', getSuiteRunResponse_suiteRunConfiguration - Suite run configuration for the test suite run.

$sel:startTime:GetSuiteRunResponse', getSuiteRunResponse_startTime - Date (in Unix epoch time) when the test suite run was started.

$sel:suiteDefinitionId:GetSuiteRun', getSuiteRunResponse_suiteDefinitionId - Suite definition Id for the test suite run.

$sel:suiteDefinitionVersion:GetSuiteRunResponse', getSuiteRunResponse_suiteDefinitionVersion - Suite definition version for the test suite run.

$sel:errorReason:GetSuiteRunResponse', getSuiteRunResponse_errorReason - Error reason for any test suite run failure.

$sel:testResult:GetSuiteRunResponse', getSuiteRunResponse_testResult - Test results for the test suite run.

$sel:endTime:GetSuiteRunResponse', getSuiteRunResponse_endTime - Date (in Unix epoch time) when the test suite run ended.

$sel:suiteRunArn:GetSuiteRunResponse', getSuiteRunResponse_suiteRunArn - The ARN of the suite run.

$sel:suiteRunId:GetSuiteRun', getSuiteRunResponse_suiteRunId - Suite run Id for the test suite run.

$sel:tags:GetSuiteRunResponse', getSuiteRunResponse_tags - The tags attached to the suite run.

$sel:httpStatus:GetSuiteRunResponse', getSuiteRunResponse_httpStatus - The response's http status code.

Response Lenses

getSuiteRunResponse_status :: Lens' GetSuiteRunResponse (Maybe SuiteRunStatus)

Status for the test suite run.

getSuiteRunResponse_suiteRunConfiguration :: Lens' GetSuiteRunResponse (Maybe SuiteRunConfiguration)

Suite run configuration for the test suite run.

getSuiteRunResponse_startTime :: Lens' GetSuiteRunResponse (Maybe UTCTime)

Date (in Unix epoch time) when the test suite run was started.

getSuiteRunResponse_suiteDefinitionId :: Lens' GetSuiteRunResponse (Maybe Text)

Suite definition Id for the test suite run.

getSuiteRunResponse_suiteDefinitionVersion :: Lens' GetSuiteRunResponse (Maybe Text)

Suite definition version for the test suite run.

getSuiteRunResponse_errorReason :: Lens' GetSuiteRunResponse (Maybe Text)

Error reason for any test suite run failure.

getSuiteRunResponse_testResult :: Lens' GetSuiteRunResponse (Maybe TestResult)

Test results for the test suite run.

getSuiteRunResponse_endTime :: Lens' GetSuiteRunResponse (Maybe UTCTime)

Date (in Unix epoch time) when the test suite run ended.

getSuiteRunResponse_suiteRunArn :: Lens' GetSuiteRunResponse (Maybe Text)

The ARN of the suite run.

getSuiteRunResponse_suiteRunId :: Lens' GetSuiteRunResponse (Maybe Text)

Suite run Id for the test suite run.

getSuiteRunResponse_tags :: Lens' GetSuiteRunResponse (Maybe (HashMap Text Text))

The tags attached to the suite run.

getSuiteRunResponse_httpStatus :: Lens' GetSuiteRunResponse Int

The response's http status code.