Amazonka.IoTDeviceAdvisor.GetSuiteRunReport

Description

Gets a report download link for a successful Device Advisor qualifying test suite run.

Creating a Request

data GetSuiteRunReport

See: newGetSuiteRunReport smart constructor.

Constructors

GetSuiteRunReport'

Fields

suiteDefinitionId :: Text

Suite definition Id of the test suite.

suiteRunId :: Text

Suite run Id of the test suite run.

newGetSuiteRunReport

Arguments

:: Text
$sel:suiteDefinitionId:GetSuiteRunReport'

-> Text
$sel:suiteRunId:GetSuiteRunReport'

-> GetSuiteRunReport

Create a value of GetSuiteRunReport with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:suiteDefinitionId:GetSuiteRunReport', getSuiteRunReport_suiteDefinitionId - Suite definition Id of the test suite.

$sel:suiteRunId:GetSuiteRunReport', getSuiteRunReport_suiteRunId - Suite run Id of the test suite run.

Request Lenses

getSuiteRunReport_suiteDefinitionId :: Lens' GetSuiteRunReport Text

Suite definition Id of the test suite.

getSuiteRunReport_suiteRunId :: Lens' GetSuiteRunReport Text

Suite run Id of the test suite run.

Destructuring the Response

data GetSuiteRunReportResponse

See: newGetSuiteRunReportResponse smart constructor.

Constructors

GetSuiteRunReportResponse'

Fields

qualificationReportDownloadUrl :: Maybe Text

Download URL of the qualification report.

httpStatus :: Int

The response's http status code.

newGetSuiteRunReportResponse

Arguments

:: Int
$sel:httpStatus:GetSuiteRunReportResponse'

-> GetSuiteRunReportResponse

Create a value of GetSuiteRunReportResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:qualificationReportDownloadUrl:GetSuiteRunReportResponse', getSuiteRunReportResponse_qualificationReportDownloadUrl - Download URL of the qualification report.

$sel:httpStatus:GetSuiteRunReportResponse', getSuiteRunReportResponse_httpStatus - The response's http status code.

Response Lenses

getSuiteRunReportResponse_qualificationReportDownloadUrl :: Lens' GetSuiteRunReportResponse (Maybe Text)

Download URL of the qualification report.

getSuiteRunReportResponse_httpStatus :: Lens' GetSuiteRunReportResponse Int

The response's http status code.