Amazonka.IoTDeviceAdvisor.StartSuiteRun

Description

Starts a Device Advisor test suite run.

Creating a Request

data StartSuiteRun

See: newStartSuiteRun smart constructor.

Constructors

StartSuiteRun'

Fields

suiteRunConfiguration :: Maybe SuiteRunConfiguration

Suite run configuration.

suiteDefinitionVersion :: Maybe Text

Suite definition version of the test suite.

tags :: Maybe (HashMap Text Text)

The tags to be attached to the suite run.

suiteDefinitionId :: Text

Suite definition Id of the test suite.

newStartSuiteRun

Arguments

:: Text
$sel:suiteDefinitionId:StartSuiteRun'

-> StartSuiteRun

Create a value of StartSuiteRun with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:suiteRunConfiguration:StartSuiteRun', startSuiteRun_suiteRunConfiguration - Suite run configuration.

$sel:suiteDefinitionVersion:StartSuiteRun', startSuiteRun_suiteDefinitionVersion - Suite definition version of the test suite.

$sel:tags:StartSuiteRun', startSuiteRun_tags - The tags to be attached to the suite run.

$sel:suiteDefinitionId:StartSuiteRun', startSuiteRun_suiteDefinitionId - Suite definition Id of the test suite.

Request Lenses

startSuiteRun_suiteRunConfiguration :: Lens' StartSuiteRun (Maybe SuiteRunConfiguration)

Suite run configuration.

startSuiteRun_suiteDefinitionVersion :: Lens' StartSuiteRun (Maybe Text)

Suite definition version of the test suite.

startSuiteRun_tags :: Lens' StartSuiteRun (Maybe (HashMap Text Text))

The tags to be attached to the suite run.

startSuiteRun_suiteDefinitionId :: Lens' StartSuiteRun Text

Suite definition Id of the test suite.

Destructuring the Response

data StartSuiteRunResponse

See: newStartSuiteRunResponse smart constructor.

Constructors

StartSuiteRunResponse'

Fields

createdAt :: Maybe POSIX

Date (in Unix epoch time) when the suite run was created.

suiteRunArn :: Maybe Text

Amazon resource name of the started suite run.

suiteRunId :: Maybe Text

Suite Run Id of the started suite run.

httpStatus :: Int

The response's http status code.

newStartSuiteRunResponse

Arguments

:: Int
$sel:httpStatus:StartSuiteRunResponse'

-> StartSuiteRunResponse

Create a value of StartSuiteRunResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:createdAt:StartSuiteRunResponse', startSuiteRunResponse_createdAt - Date (in Unix epoch time) when the suite run was created.

$sel:suiteRunArn:StartSuiteRunResponse', startSuiteRunResponse_suiteRunArn - Amazon resource name of the started suite run.

$sel:suiteRunId:StartSuiteRunResponse', startSuiteRunResponse_suiteRunId - Suite Run Id of the started suite run.

$sel:httpStatus:StartSuiteRunResponse', startSuiteRunResponse_httpStatus - The response's http status code.

Response Lenses

startSuiteRunResponse_createdAt :: Lens' StartSuiteRunResponse (Maybe UTCTime)

Date (in Unix epoch time) when the suite run was created.

startSuiteRunResponse_suiteRunArn :: Lens' StartSuiteRunResponse (Maybe Text)

Amazon resource name of the started suite run.

startSuiteRunResponse_suiteRunId :: Lens' StartSuiteRunResponse (Maybe Text)

Suite Run Id of the started suite run.

startSuiteRunResponse_httpStatus :: Lens' StartSuiteRunResponse Int

The response's http status code.