Amazonka.IoTEvents.CreateAlarmModel

Description

Creates an alarm model to monitor an AWS IoT Events input attribute. You can use the alarm to get notified when the value is outside a specified range. For more information, see Create an alarm model in the AWS IoT Events Developer Guide.

Creating a Request

data CreateAlarmModel

See: newCreateAlarmModel smart constructor.

Constructors

CreateAlarmModel'

Fields

severity :: Maybe Natural

A non-negative integer that reflects the severity level of the alarm.

alarmNotification :: Maybe AlarmNotification

Contains information about one or more notification actions.

key :: Maybe Text

An input attribute used as a key to create an alarm. AWS IoT Events routes inputs associated with this key to the alarm.

alarmModelDescription :: Maybe Text

A description that tells you what the alarm model detects.

alarmEventActions :: Maybe AlarmEventActions

Contains information about one or more alarm actions.

alarmCapabilities :: Maybe AlarmCapabilities

Contains the configuration information of alarm state changes.

tags :: Maybe [Tag]

A list of key-value pairs that contain metadata for the alarm model. The tags help you manage the alarm model. For more information, see Tagging your AWS IoT Events resources in the AWS IoT Events Developer Guide.

You can create up to 50 tags for one alarm model.

alarmModelName :: Text

A unique name that helps you identify the alarm model. You can't change this name after you create the alarm model.

roleArn :: Text

The ARN of the IAM role that allows the alarm to perform actions and access AWS resources. For more information, see Amazon Resource Names (ARNs) in the AWS General Reference.

alarmRule :: AlarmRule

Defines when your alarm is invoked.

newCreateAlarmModel

Arguments

:: Text
$sel:alarmModelName:CreateAlarmModel'

-> Text
$sel:roleArn:CreateAlarmModel'

-> AlarmRule
$sel:alarmRule:CreateAlarmModel'

-> CreateAlarmModel

Create a value of CreateAlarmModel with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:severity:CreateAlarmModel', createAlarmModel_severity - A non-negative integer that reflects the severity level of the alarm.

$sel:alarmNotification:CreateAlarmModel', createAlarmModel_alarmNotification - Contains information about one or more notification actions.

$sel:key:CreateAlarmModel', createAlarmModel_key - An input attribute used as a key to create an alarm. AWS IoT Events routes inputs associated with this key to the alarm.

$sel:alarmModelDescription:CreateAlarmModel', createAlarmModel_alarmModelDescription - A description that tells you what the alarm model detects.

$sel:alarmEventActions:CreateAlarmModel', createAlarmModel_alarmEventActions - Contains information about one or more alarm actions.

$sel:alarmCapabilities:CreateAlarmModel', createAlarmModel_alarmCapabilities - Contains the configuration information of alarm state changes.

$sel:tags:CreateAlarmModel', createAlarmModel_tags - A list of key-value pairs that contain metadata for the alarm model. The tags help you manage the alarm model. For more information, see Tagging your AWS IoT Events resources in the AWS IoT Events Developer Guide.

You can create up to 50 tags for one alarm model.

$sel:alarmModelName:CreateAlarmModel', createAlarmModel_alarmModelName - A unique name that helps you identify the alarm model. You can't change this name after you create the alarm model.

$sel:roleArn:CreateAlarmModel', createAlarmModel_roleArn - The ARN of the IAM role that allows the alarm to perform actions and access AWS resources. For more information, see Amazon Resource Names (ARNs) in the AWS General Reference.

$sel:alarmRule:CreateAlarmModel', createAlarmModel_alarmRule - Defines when your alarm is invoked.

Request Lenses

createAlarmModel_severity :: Lens' CreateAlarmModel (Maybe Natural)

A non-negative integer that reflects the severity level of the alarm.

createAlarmModel_alarmNotification :: Lens' CreateAlarmModel (Maybe AlarmNotification)

Contains information about one or more notification actions.

createAlarmModel_key :: Lens' CreateAlarmModel (Maybe Text)

An input attribute used as a key to create an alarm. AWS IoT Events routes inputs associated with this key to the alarm.

createAlarmModel_alarmModelDescription :: Lens' CreateAlarmModel (Maybe Text)

A description that tells you what the alarm model detects.

createAlarmModel_alarmEventActions :: Lens' CreateAlarmModel (Maybe AlarmEventActions)

Contains information about one or more alarm actions.

createAlarmModel_alarmCapabilities :: Lens' CreateAlarmModel (Maybe AlarmCapabilities)

Contains the configuration information of alarm state changes.

createAlarmModel_tags :: Lens' CreateAlarmModel (Maybe [Tag])

A list of key-value pairs that contain metadata for the alarm model. The tags help you manage the alarm model. For more information, see Tagging your AWS IoT Events resources in the AWS IoT Events Developer Guide.

You can create up to 50 tags for one alarm model.

createAlarmModel_alarmModelName :: Lens' CreateAlarmModel Text

A unique name that helps you identify the alarm model. You can't change this name after you create the alarm model.

createAlarmModel_roleArn :: Lens' CreateAlarmModel Text

The ARN of the IAM role that allows the alarm to perform actions and access AWS resources. For more information, see Amazon Resource Names (ARNs) in the AWS General Reference.

createAlarmModel_alarmRule :: Lens' CreateAlarmModel AlarmRule

Defines when your alarm is invoked.

Destructuring the Response

data CreateAlarmModelResponse

See: newCreateAlarmModelResponse smart constructor.

Constructors

CreateAlarmModelResponse'

Fields

creationTime :: Maybe POSIX

The time the alarm model was created, in the Unix epoch format.

status :: Maybe AlarmModelVersionStatus

The status of the alarm model. The status can be one of the following values:

ACTIVE - The alarm model is active and it's ready to evaluate data.

ACTIVATING - AWS IoT Events is activating your alarm model. Activating an alarm model can take up to a few minutes.

INACTIVE - The alarm model is inactive, so it isn't ready to evaluate data. Check your alarm model information and update the alarm model.

FAILED - You couldn't create or update the alarm model. Check your alarm model information and try again.

lastUpdateTime :: Maybe POSIX

The time the alarm model was last updated, in the Unix epoch format.

alarmModelVersion :: Maybe Text

The version of the alarm model.

alarmModelArn :: Maybe Text

The ARN of the alarm model. For more information, see Amazon Resource Names (ARNs) in the AWS General Reference.

httpStatus :: Int

The response's http status code.

newCreateAlarmModelResponse

Arguments

:: Int
$sel:httpStatus:CreateAlarmModelResponse'

-> CreateAlarmModelResponse

Create a value of CreateAlarmModelResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:creationTime:CreateAlarmModelResponse', createAlarmModelResponse_creationTime - The time the alarm model was created, in the Unix epoch format.

$sel:status:CreateAlarmModelResponse', createAlarmModelResponse_status - The status of the alarm model. The status can be one of the following values:

ACTIVE - The alarm model is active and it's ready to evaluate data.

ACTIVATING - AWS IoT Events is activating your alarm model. Activating an alarm model can take up to a few minutes.

INACTIVE - The alarm model is inactive, so it isn't ready to evaluate data. Check your alarm model information and update the alarm model.

FAILED - You couldn't create or update the alarm model. Check your alarm model information and try again.

$sel:lastUpdateTime:CreateAlarmModelResponse', createAlarmModelResponse_lastUpdateTime - The time the alarm model was last updated, in the Unix epoch format.

$sel:alarmModelVersion:CreateAlarmModelResponse', createAlarmModelResponse_alarmModelVersion - The version of the alarm model.

$sel:alarmModelArn:CreateAlarmModelResponse', createAlarmModelResponse_alarmModelArn - The ARN of the alarm model. For more information, see Amazon Resource Names (ARNs) in the AWS General Reference.

$sel:httpStatus:CreateAlarmModelResponse', createAlarmModelResponse_httpStatus - The response's http status code.

Response Lenses

createAlarmModelResponse_creationTime :: Lens' CreateAlarmModelResponse (Maybe UTCTime)

The time the alarm model was created, in the Unix epoch format.

createAlarmModelResponse_status :: Lens' CreateAlarmModelResponse (Maybe AlarmModelVersionStatus)

The status of the alarm model. The status can be one of the following values:

ACTIVE - The alarm model is active and it's ready to evaluate data.

ACTIVATING - AWS IoT Events is activating your alarm model. Activating an alarm model can take up to a few minutes.

INACTIVE - The alarm model is inactive, so it isn't ready to evaluate data. Check your alarm model information and update the alarm model.

FAILED - You couldn't create or update the alarm model. Check your alarm model information and try again.

createAlarmModelResponse_lastUpdateTime :: Lens' CreateAlarmModelResponse (Maybe UTCTime)

The time the alarm model was last updated, in the Unix epoch format.

createAlarmModelResponse_alarmModelVersion :: Lens' CreateAlarmModelResponse (Maybe Text)

The version of the alarm model.

createAlarmModelResponse_alarmModelArn :: Lens' CreateAlarmModelResponse (Maybe Text)

The ARN of the alarm model. For more information, see Amazon Resource Names (ARNs) in the AWS General Reference.

createAlarmModelResponse_httpStatus :: Lens' CreateAlarmModelResponse Int

The response's http status code.