Amazonka.IoTEvents.DeleteInput

Description

Deletes an input.

Creating a Request

data DeleteInput

See: newDeleteInput smart constructor.

Constructors

DeleteInput'

Fields

inputName :: Text

The name of the input to delete.

newDeleteInput

Arguments

:: Text
$sel:inputName:DeleteInput'

-> DeleteInput

Create a value of DeleteInput with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:inputName:DeleteInput', deleteInput_inputName - The name of the input to delete.

Request Lenses

deleteInput_inputName :: Lens' DeleteInput Text

The name of the input to delete.

Destructuring the Response

data DeleteInputResponse

See: newDeleteInputResponse smart constructor.

Constructors

DeleteInputResponse'

Fields

httpStatus :: Int

The response's http status code.

newDeleteInputResponse

Arguments

:: Int
$sel:httpStatus:DeleteInputResponse'

-> DeleteInputResponse

Create a value of DeleteInputResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:httpStatus:DeleteInputResponse', deleteInputResponse_httpStatus - The response's http status code.

Response Lenses

deleteInputResponse_httpStatus :: Lens' DeleteInputResponse Int

The response's http status code.