Amazonka.IoTEvents.DescribeAlarmModel

Description

Retrieves information about an alarm model. If you don't specify a value for the alarmModelVersion parameter, the latest version is returned.

Creating a Request

data DescribeAlarmModel

See: newDescribeAlarmModel smart constructor.

Constructors

DescribeAlarmModel'

Fields

alarmModelVersion :: Maybe Text

The version of the alarm model.

alarmModelName :: Text

The name of the alarm model.

newDescribeAlarmModel

Arguments

:: Text
$sel:alarmModelName:DescribeAlarmModel'

-> DescribeAlarmModel

Create a value of DescribeAlarmModel with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:alarmModelVersion:DescribeAlarmModel', describeAlarmModel_alarmModelVersion - The version of the alarm model.

$sel:alarmModelName:DescribeAlarmModel', describeAlarmModel_alarmModelName - The name of the alarm model.

Request Lenses

describeAlarmModel_alarmModelVersion :: Lens' DescribeAlarmModel (Maybe Text)

The version of the alarm model.

describeAlarmModel_alarmModelName :: Lens' DescribeAlarmModel Text

The name of the alarm model.

Destructuring the Response

data DescribeAlarmModelResponse

See: newDescribeAlarmModelResponse smart constructor.

Constructors

DescribeAlarmModelResponse'

Fields

creationTime :: Maybe POSIX

The time the alarm model was created, in the Unix epoch format.

status :: Maybe AlarmModelVersionStatus

The status of the alarm model. The status can be one of the following values:

ACTIVE - The alarm model is active and it's ready to evaluate data.

ACTIVATING - AWS IoT Events is activating your alarm model. Activating an alarm model can take up to a few minutes.

INACTIVE - The alarm model is inactive, so it isn't ready to evaluate data. Check your alarm model information and update the alarm model.

FAILED - You couldn't create or update the alarm model. Check your alarm model information and try again.

alarmRule :: Maybe AlarmRule

Defines when your alarm is invoked.

alarmModelName :: Maybe Text

The name of the alarm model.

severity :: Maybe Natural

A non-negative integer that reflects the severity level of the alarm.

alarmNotification :: Maybe AlarmNotification

Contains information about one or more notification actions.

key :: Maybe Text

An input attribute used as a key to create an alarm. AWS IoT Events routes inputs associated with this key to the alarm.

statusMessage :: Maybe Text

Contains information about the status of the alarm model.

alarmModelDescription :: Maybe Text

The description of the alarm model.

alarmEventActions :: Maybe AlarmEventActions

Contains information about one or more alarm actions.

alarmCapabilities :: Maybe AlarmCapabilities

Contains the configuration information of alarm state changes.

lastUpdateTime :: Maybe POSIX

The time the alarm model was last updated, in the Unix epoch format.

alarmModelVersion :: Maybe Text

The version of the alarm model.

alarmModelArn :: Maybe Text

The ARN of the alarm model. For more information, see Amazon Resource Names (ARNs) in the AWS General Reference.

roleArn :: Maybe Text

The ARN of the IAM role that allows the alarm to perform actions and access AWS resources. For more information, see Amazon Resource Names (ARNs) in the AWS General Reference.

httpStatus :: Int

The response's http status code.

newDescribeAlarmModelResponse

Arguments

:: Int
$sel:httpStatus:DescribeAlarmModelResponse'

-> DescribeAlarmModelResponse

Create a value of DescribeAlarmModelResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:creationTime:DescribeAlarmModelResponse', describeAlarmModelResponse_creationTime - The time the alarm model was created, in the Unix epoch format.

$sel:status:DescribeAlarmModelResponse', describeAlarmModelResponse_status - The status of the alarm model. The status can be one of the following values:

ACTIVE - The alarm model is active and it's ready to evaluate data.

ACTIVATING - AWS IoT Events is activating your alarm model. Activating an alarm model can take up to a few minutes.

INACTIVE - The alarm model is inactive, so it isn't ready to evaluate data. Check your alarm model information and update the alarm model.

FAILED - You couldn't create or update the alarm model. Check your alarm model information and try again.

$sel:alarmRule:DescribeAlarmModelResponse', describeAlarmModelResponse_alarmRule - Defines when your alarm is invoked.

$sel:alarmModelName:DescribeAlarmModel', describeAlarmModelResponse_alarmModelName - The name of the alarm model.

$sel:severity:DescribeAlarmModelResponse', describeAlarmModelResponse_severity - A non-negative integer that reflects the severity level of the alarm.

$sel:alarmNotification:DescribeAlarmModelResponse', describeAlarmModelResponse_alarmNotification - Contains information about one or more notification actions.

$sel:key:DescribeAlarmModelResponse', describeAlarmModelResponse_key - An input attribute used as a key to create an alarm. AWS IoT Events routes inputs associated with this key to the alarm.

$sel:statusMessage:DescribeAlarmModelResponse', describeAlarmModelResponse_statusMessage - Contains information about the status of the alarm model.

$sel:alarmModelDescription:DescribeAlarmModelResponse', describeAlarmModelResponse_alarmModelDescription - The description of the alarm model.

$sel:alarmEventActions:DescribeAlarmModelResponse', describeAlarmModelResponse_alarmEventActions - Contains information about one or more alarm actions.

$sel:alarmCapabilities:DescribeAlarmModelResponse', describeAlarmModelResponse_alarmCapabilities - Contains the configuration information of alarm state changes.

$sel:lastUpdateTime:DescribeAlarmModelResponse', describeAlarmModelResponse_lastUpdateTime - The time the alarm model was last updated, in the Unix epoch format.

$sel:alarmModelVersion:DescribeAlarmModel', describeAlarmModelResponse_alarmModelVersion - The version of the alarm model.

$sel:alarmModelArn:DescribeAlarmModelResponse', describeAlarmModelResponse_alarmModelArn - The ARN of the alarm model. For more information, see Amazon Resource Names (ARNs) in the AWS General Reference.

$sel:roleArn:DescribeAlarmModelResponse', describeAlarmModelResponse_roleArn - The ARN of the IAM role that allows the alarm to perform actions and access AWS resources. For more information, see Amazon Resource Names (ARNs) in the AWS General Reference.

$sel:httpStatus:DescribeAlarmModelResponse', describeAlarmModelResponse_httpStatus - The response's http status code.

Response Lenses

describeAlarmModelResponse_creationTime :: Lens' DescribeAlarmModelResponse (Maybe UTCTime)

The time the alarm model was created, in the Unix epoch format.

describeAlarmModelResponse_status :: Lens' DescribeAlarmModelResponse (Maybe AlarmModelVersionStatus)

The status of the alarm model. The status can be one of the following values:

ACTIVE - The alarm model is active and it's ready to evaluate data.

ACTIVATING - AWS IoT Events is activating your alarm model. Activating an alarm model can take up to a few minutes.

INACTIVE - The alarm model is inactive, so it isn't ready to evaluate data. Check your alarm model information and update the alarm model.

FAILED - You couldn't create or update the alarm model. Check your alarm model information and try again.

describeAlarmModelResponse_alarmRule :: Lens' DescribeAlarmModelResponse (Maybe AlarmRule)

Defines when your alarm is invoked.

describeAlarmModelResponse_alarmModelName :: Lens' DescribeAlarmModelResponse (Maybe Text)

The name of the alarm model.

describeAlarmModelResponse_severity :: Lens' DescribeAlarmModelResponse (Maybe Natural)

A non-negative integer that reflects the severity level of the alarm.

describeAlarmModelResponse_alarmNotification :: Lens' DescribeAlarmModelResponse (Maybe AlarmNotification)

Contains information about one or more notification actions.

describeAlarmModelResponse_key :: Lens' DescribeAlarmModelResponse (Maybe Text)

An input attribute used as a key to create an alarm. AWS IoT Events routes inputs associated with this key to the alarm.

describeAlarmModelResponse_statusMessage :: Lens' DescribeAlarmModelResponse (Maybe Text)

Contains information about the status of the alarm model.

describeAlarmModelResponse_alarmModelDescription :: Lens' DescribeAlarmModelResponse (Maybe Text)

The description of the alarm model.

describeAlarmModelResponse_alarmEventActions :: Lens' DescribeAlarmModelResponse (Maybe AlarmEventActions)

Contains information about one or more alarm actions.

describeAlarmModelResponse_alarmCapabilities :: Lens' DescribeAlarmModelResponse (Maybe AlarmCapabilities)

Contains the configuration information of alarm state changes.

describeAlarmModelResponse_lastUpdateTime :: Lens' DescribeAlarmModelResponse (Maybe UTCTime)

The time the alarm model was last updated, in the Unix epoch format.

describeAlarmModelResponse_alarmModelVersion :: Lens' DescribeAlarmModelResponse (Maybe Text)

The version of the alarm model.

describeAlarmModelResponse_alarmModelArn :: Lens' DescribeAlarmModelResponse (Maybe Text)

The ARN of the alarm model. For more information, see Amazon Resource Names (ARNs) in the AWS General Reference.

describeAlarmModelResponse_roleArn :: Lens' DescribeAlarmModelResponse (Maybe Text)

The ARN of the IAM role that allows the alarm to perform actions and access AWS resources. For more information, see Amazon Resource Names (ARNs) in the AWS General Reference.

describeAlarmModelResponse_httpStatus :: Lens' DescribeAlarmModelResponse Int

The response's http status code.