Amazonka.IoTEvents.Lens

Description

Operations

ListInputs

listInputs_nextToken :: Lens' ListInputs (Maybe Text)

The token that you can use to return the next set of results.

listInputs_maxResults :: Lens' ListInputs (Maybe Natural)

The maximum number of results to be returned per request.

listInputsResponse_inputSummaries :: Lens' ListInputsResponse (Maybe [InputSummary])

Summary information about the inputs.

listInputsResponse_nextToken :: Lens' ListInputsResponse (Maybe Text)

The token that you can use to return the next set of results, or null if there are no more results.

listInputsResponse_httpStatus :: Lens' ListInputsResponse Int

The response's http status code.

StartDetectorModelAnalysis

startDetectorModelAnalysis_detectorModelDefinition :: Lens' StartDetectorModelAnalysis DetectorModelDefinition

Undocumented member.

startDetectorModelAnalysisResponse_analysisId :: Lens' StartDetectorModelAnalysisResponse (Maybe Text)

The ID that you can use to retrieve the analysis result.

startDetectorModelAnalysisResponse_httpStatus :: Lens' StartDetectorModelAnalysisResponse Int

The response's http status code.

PutLoggingOptions

putLoggingOptions_loggingOptions :: Lens' PutLoggingOptions LoggingOptions

The new values of the AWS IoT Events logging options.

DescribeDetectorModelAnalysis

describeDetectorModelAnalysis_analysisId :: Lens' DescribeDetectorModelAnalysis Text

The ID of the analysis result that you want to retrieve.

describeDetectorModelAnalysisResponse_status :: Lens' DescribeDetectorModelAnalysisResponse (Maybe AnalysisStatus)

The status of the analysis activity. The status can be one of the following values:

RUNNING - AWS IoT Events is analyzing your detector model. This process can take several minutes to complete.

COMPLETE - AWS IoT Events finished analyzing your detector model.

FAILED - AWS IoT Events couldn't analyze your detector model. Try again later.

describeDetectorModelAnalysisResponse_httpStatus :: Lens' DescribeDetectorModelAnalysisResponse Int

The response's http status code.

CreateInput

createInput_inputDescription :: Lens' CreateInput (Maybe Text)

A brief description of the input.

createInput_tags :: Lens' CreateInput (Maybe [Tag])

Metadata that can be used to manage the input.

createInput_inputName :: Lens' CreateInput Text

The name you want to give to the input.

createInput_inputDefinition :: Lens' CreateInput InputDefinition

The definition of the input.

createInputResponse_inputConfiguration :: Lens' CreateInputResponse (Maybe InputConfiguration)

Information about the configuration of the input.

createInputResponse_httpStatus :: Lens' CreateInputResponse Int

The response's http status code.

ListTagsForResource

listTagsForResource_resourceArn :: Lens' ListTagsForResource Text

The ARN of the resource.

listTagsForResourceResponse_tags :: Lens' ListTagsForResourceResponse (Maybe [Tag])

The list of tags assigned to the resource.

listTagsForResourceResponse_httpStatus :: Lens' ListTagsForResourceResponse Int

The response's http status code.

ListAlarmModels

listAlarmModels_nextToken :: Lens' ListAlarmModels (Maybe Text)

The token that you can use to return the next set of results.

listAlarmModels_maxResults :: Lens' ListAlarmModels (Maybe Natural)

The maximum number of results to be returned per request.

listAlarmModelsResponse_nextToken :: Lens' ListAlarmModelsResponse (Maybe Text)

The token that you can use to return the next set of results, or null if there are no more results.

listAlarmModelsResponse_alarmModelSummaries :: Lens' ListAlarmModelsResponse (Maybe [AlarmModelSummary])

A list that summarizes each alarm model.

listAlarmModelsResponse_httpStatus :: Lens' ListAlarmModelsResponse Int

The response's http status code.

DeleteAlarmModel

deleteAlarmModel_alarmModelName :: Lens' DeleteAlarmModel Text

The name of the alarm model.

deleteAlarmModelResponse_httpStatus :: Lens' DeleteAlarmModelResponse Int

The response's http status code.

UpdateAlarmModel

updateAlarmModel_severity :: Lens' UpdateAlarmModel (Maybe Natural)

A non-negative integer that reflects the severity level of the alarm.

updateAlarmModel_alarmNotification :: Lens' UpdateAlarmModel (Maybe AlarmNotification)

Contains information about one or more notification actions.

updateAlarmModel_alarmModelDescription :: Lens' UpdateAlarmModel (Maybe Text)

The description of the alarm model.

updateAlarmModel_alarmEventActions :: Lens' UpdateAlarmModel (Maybe AlarmEventActions)

Contains information about one or more alarm actions.

updateAlarmModel_alarmCapabilities :: Lens' UpdateAlarmModel (Maybe AlarmCapabilities)

Contains the configuration information of alarm state changes.

updateAlarmModel_alarmModelName :: Lens' UpdateAlarmModel Text

The name of the alarm model.

updateAlarmModel_roleArn :: Lens' UpdateAlarmModel Text

The ARN of the IAM role that allows the alarm to perform actions and access AWS resources. For more information, see Amazon Resource Names (ARNs) in the AWS General Reference.

updateAlarmModel_alarmRule :: Lens' UpdateAlarmModel AlarmRule

Defines when your alarm is invoked.

updateAlarmModelResponse_creationTime :: Lens' UpdateAlarmModelResponse (Maybe UTCTime)

The time the alarm model was created, in the Unix epoch format.

updateAlarmModelResponse_status :: Lens' UpdateAlarmModelResponse (Maybe AlarmModelVersionStatus)

The status of the alarm model. The status can be one of the following values:

ACTIVE - The alarm model is active and it's ready to evaluate data.

ACTIVATING - AWS IoT Events is activating your alarm model. Activating an alarm model can take up to a few minutes.

INACTIVE - The alarm model is inactive, so it isn't ready to evaluate data. Check your alarm model information and update the alarm model.

FAILED - You couldn't create or update the alarm model. Check your alarm model information and try again.

updateAlarmModelResponse_lastUpdateTime :: Lens' UpdateAlarmModelResponse (Maybe UTCTime)

The time the alarm model was last updated, in the Unix epoch format.

updateAlarmModelResponse_alarmModelVersion :: Lens' UpdateAlarmModelResponse (Maybe Text)

The version of the alarm model.

updateAlarmModelResponse_alarmModelArn :: Lens' UpdateAlarmModelResponse (Maybe Text)

The ARN of the alarm model. For more information, see Amazon Resource Names (ARNs) in the AWS General Reference.

updateAlarmModelResponse_httpStatus :: Lens' UpdateAlarmModelResponse Int

The response's http status code.

CreateAlarmModel

createAlarmModel_severity :: Lens' CreateAlarmModel (Maybe Natural)

A non-negative integer that reflects the severity level of the alarm.

createAlarmModel_alarmNotification :: Lens' CreateAlarmModel (Maybe AlarmNotification)

Contains information about one or more notification actions.

createAlarmModel_key :: Lens' CreateAlarmModel (Maybe Text)

An input attribute used as a key to create an alarm. AWS IoT Events routes inputs associated with this key to the alarm.

createAlarmModel_alarmModelDescription :: Lens' CreateAlarmModel (Maybe Text)

A description that tells you what the alarm model detects.

createAlarmModel_alarmEventActions :: Lens' CreateAlarmModel (Maybe AlarmEventActions)

Contains information about one or more alarm actions.

createAlarmModel_alarmCapabilities :: Lens' CreateAlarmModel (Maybe AlarmCapabilities)

Contains the configuration information of alarm state changes.

createAlarmModel_tags :: Lens' CreateAlarmModel (Maybe [Tag])

A list of key-value pairs that contain metadata for the alarm model. The tags help you manage the alarm model. For more information, see Tagging your AWS IoT Events resources in the AWS IoT Events Developer Guide.

You can create up to 50 tags for one alarm model.

createAlarmModel_alarmModelName :: Lens' CreateAlarmModel Text

A unique name that helps you identify the alarm model. You can't change this name after you create the alarm model.

createAlarmModel_roleArn :: Lens' CreateAlarmModel Text

The ARN of the IAM role that allows the alarm to perform actions and access AWS resources. For more information, see Amazon Resource Names (ARNs) in the AWS General Reference.

createAlarmModel_alarmRule :: Lens' CreateAlarmModel AlarmRule

Defines when your alarm is invoked.

createAlarmModelResponse_creationTime :: Lens' CreateAlarmModelResponse (Maybe UTCTime)

The time the alarm model was created, in the Unix epoch format.

createAlarmModelResponse_status :: Lens' CreateAlarmModelResponse (Maybe AlarmModelVersionStatus)

The status of the alarm model. The status can be one of the following values:

ACTIVE - The alarm model is active and it's ready to evaluate data.

ACTIVATING - AWS IoT Events is activating your alarm model. Activating an alarm model can take up to a few minutes.

INACTIVE - The alarm model is inactive, so it isn't ready to evaluate data. Check your alarm model information and update the alarm model.

FAILED - You couldn't create or update the alarm model. Check your alarm model information and try again.

createAlarmModelResponse_lastUpdateTime :: Lens' CreateAlarmModelResponse (Maybe UTCTime)

The time the alarm model was last updated, in the Unix epoch format.

createAlarmModelResponse_alarmModelVersion :: Lens' CreateAlarmModelResponse (Maybe Text)

The version of the alarm model.

createAlarmModelResponse_alarmModelArn :: Lens' CreateAlarmModelResponse (Maybe Text)

The ARN of the alarm model. For more information, see Amazon Resource Names (ARNs) in the AWS General Reference.

createAlarmModelResponse_httpStatus :: Lens' CreateAlarmModelResponse Int

The response's http status code.

GetDetectorModelAnalysisResults

getDetectorModelAnalysisResults_nextToken :: Lens' GetDetectorModelAnalysisResults (Maybe Text)

The token that you can use to return the next set of results.

getDetectorModelAnalysisResults_maxResults :: Lens' GetDetectorModelAnalysisResults (Maybe Int)

The maximum number of results to be returned per request.

getDetectorModelAnalysisResults_analysisId :: Lens' GetDetectorModelAnalysisResults Text

The ID of the analysis result that you want to retrieve.

getDetectorModelAnalysisResultsResponse_nextToken :: Lens' GetDetectorModelAnalysisResultsResponse (Maybe Text)

The token that you can use to return the next set of results, or null if there are no more results.

getDetectorModelAnalysisResultsResponse_analysisResults :: Lens' GetDetectorModelAnalysisResultsResponse (Maybe [AnalysisResult])

Contains information about one or more analysis results.

getDetectorModelAnalysisResultsResponse_httpStatus :: Lens' GetDetectorModelAnalysisResultsResponse Int

The response's http status code.

ListDetectorModelVersions

listDetectorModelVersions_nextToken :: Lens' ListDetectorModelVersions (Maybe Text)

The token that you can use to return the next set of results.

listDetectorModelVersions_maxResults :: Lens' ListDetectorModelVersions (Maybe Natural)

The maximum number of results to be returned per request.

listDetectorModelVersions_detectorModelName :: Lens' ListDetectorModelVersions Text

The name of the detector model whose versions are returned.

listDetectorModelVersionsResponse_nextToken :: Lens' ListDetectorModelVersionsResponse (Maybe Text)

The token that you can use to return the next set of results, or null if there are no more results.

listDetectorModelVersionsResponse_detectorModelVersionSummaries :: Lens' ListDetectorModelVersionsResponse (Maybe [DetectorModelVersionSummary])

Summary information about the detector model versions.

listDetectorModelVersionsResponse_httpStatus :: Lens' ListDetectorModelVersionsResponse Int

The response's http status code.

DescribeAlarmModel

describeAlarmModel_alarmModelVersion :: Lens' DescribeAlarmModel (Maybe Text)

The version of the alarm model.

describeAlarmModel_alarmModelName :: Lens' DescribeAlarmModel Text

The name of the alarm model.

describeAlarmModelResponse_creationTime :: Lens' DescribeAlarmModelResponse (Maybe UTCTime)

The time the alarm model was created, in the Unix epoch format.

describeAlarmModelResponse_status :: Lens' DescribeAlarmModelResponse (Maybe AlarmModelVersionStatus)

The status of the alarm model. The status can be one of the following values:

ACTIVE - The alarm model is active and it's ready to evaluate data.

ACTIVATING - AWS IoT Events is activating your alarm model. Activating an alarm model can take up to a few minutes.

INACTIVE - The alarm model is inactive, so it isn't ready to evaluate data. Check your alarm model information and update the alarm model.

FAILED - You couldn't create or update the alarm model. Check your alarm model information and try again.

describeAlarmModelResponse_alarmRule :: Lens' DescribeAlarmModelResponse (Maybe AlarmRule)

Defines when your alarm is invoked.

describeAlarmModelResponse_alarmModelName :: Lens' DescribeAlarmModelResponse (Maybe Text)

The name of the alarm model.

describeAlarmModelResponse_severity :: Lens' DescribeAlarmModelResponse (Maybe Natural)

A non-negative integer that reflects the severity level of the alarm.

describeAlarmModelResponse_alarmNotification :: Lens' DescribeAlarmModelResponse (Maybe AlarmNotification)

Contains information about one or more notification actions.

describeAlarmModelResponse_key :: Lens' DescribeAlarmModelResponse (Maybe Text)

An input attribute used as a key to create an alarm. AWS IoT Events routes inputs associated with this key to the alarm.

describeAlarmModelResponse_statusMessage :: Lens' DescribeAlarmModelResponse (Maybe Text)

Contains information about the status of the alarm model.

describeAlarmModelResponse_alarmModelDescription :: Lens' DescribeAlarmModelResponse (Maybe Text)

The description of the alarm model.

describeAlarmModelResponse_alarmEventActions :: Lens' DescribeAlarmModelResponse (Maybe AlarmEventActions)

Contains information about one or more alarm actions.

describeAlarmModelResponse_alarmCapabilities :: Lens' DescribeAlarmModelResponse (Maybe AlarmCapabilities)

Contains the configuration information of alarm state changes.

describeAlarmModelResponse_lastUpdateTime :: Lens' DescribeAlarmModelResponse (Maybe UTCTime)

The time the alarm model was last updated, in the Unix epoch format.

describeAlarmModelResponse_alarmModelVersion :: Lens' DescribeAlarmModelResponse (Maybe Text)

The version of the alarm model.

describeAlarmModelResponse_alarmModelArn :: Lens' DescribeAlarmModelResponse (Maybe Text)

The ARN of the alarm model. For more information, see Amazon Resource Names (ARNs) in the AWS General Reference.

describeAlarmModelResponse_roleArn :: Lens' DescribeAlarmModelResponse (Maybe Text)

The ARN of the IAM role that allows the alarm to perform actions and access AWS resources. For more information, see Amazon Resource Names (ARNs) in the AWS General Reference.

describeAlarmModelResponse_httpStatus :: Lens' DescribeAlarmModelResponse Int

The response's http status code.

CreateDetectorModel

createDetectorModel_detectorModelDescription :: Lens' CreateDetectorModel (Maybe Text)

A brief description of the detector model.

createDetectorModel_key :: Lens' CreateDetectorModel (Maybe Text)

The input attribute key used to identify a device or system to create a detector (an instance of the detector model) and then to route each input received to the appropriate detector (instance). This parameter uses a JSON-path expression in the message payload of each input to specify the attribute-value pair that is used to identify the device associated with the input.

createDetectorModel_evaluationMethod :: Lens' CreateDetectorModel (Maybe EvaluationMethod)

Information about the order in which events are evaluated and how actions are executed.

createDetectorModel_tags :: Lens' CreateDetectorModel (Maybe [Tag])

Metadata that can be used to manage the detector model.

createDetectorModel_detectorModelName :: Lens' CreateDetectorModel Text

The name of the detector model.

createDetectorModel_detectorModelDefinition :: Lens' CreateDetectorModel DetectorModelDefinition

Information that defines how the detectors operate.

createDetectorModel_roleArn :: Lens' CreateDetectorModel Text

The ARN of the role that grants permission to AWS IoT Events to perform its operations.

createDetectorModelResponse_detectorModelConfiguration :: Lens' CreateDetectorModelResponse (Maybe DetectorModelConfiguration)

Information about how the detector model is configured.

createDetectorModelResponse_httpStatus :: Lens' CreateDetectorModelResponse Int

The response's http status code.

ListDetectorModels

listDetectorModels_nextToken :: Lens' ListDetectorModels (Maybe Text)

The token that you can use to return the next set of results.

listDetectorModels_maxResults :: Lens' ListDetectorModels (Maybe Natural)

The maximum number of results to be returned per request.

listDetectorModelsResponse_nextToken :: Lens' ListDetectorModelsResponse (Maybe Text)

The token that you can use to return the next set of results, or null if there are no more results.

listDetectorModelsResponse_detectorModelSummaries :: Lens' ListDetectorModelsResponse (Maybe [DetectorModelSummary])

Summary information about the detector models.

listDetectorModelsResponse_httpStatus :: Lens' ListDetectorModelsResponse Int

The response's http status code.

UpdateDetectorModel

updateDetectorModel_detectorModelDescription :: Lens' UpdateDetectorModel (Maybe Text)

A brief description of the detector model.

updateDetectorModel_evaluationMethod :: Lens' UpdateDetectorModel (Maybe EvaluationMethod)

Information about the order in which events are evaluated and how actions are executed.

updateDetectorModel_detectorModelName :: Lens' UpdateDetectorModel Text

The name of the detector model that is updated.

updateDetectorModel_detectorModelDefinition :: Lens' UpdateDetectorModel DetectorModelDefinition

Information that defines how a detector operates.

updateDetectorModel_roleArn :: Lens' UpdateDetectorModel Text

The ARN of the role that grants permission to AWS IoT Events to perform its operations.

updateDetectorModelResponse_detectorModelConfiguration :: Lens' UpdateDetectorModelResponse (Maybe DetectorModelConfiguration)

Information about how the detector model is configured.

updateDetectorModelResponse_httpStatus :: Lens' UpdateDetectorModelResponse Int

The response's http status code.

DeleteDetectorModel

deleteDetectorModel_detectorModelName :: Lens' DeleteDetectorModel Text

The name of the detector model to be deleted.

deleteDetectorModelResponse_httpStatus :: Lens' DeleteDetectorModelResponse Int

The response's http status code.

DeleteInput

deleteInput_inputName :: Lens' DeleteInput Text

The name of the input to delete.

deleteInputResponse_httpStatus :: Lens' DeleteInputResponse Int

The response's http status code.

UpdateInput

updateInput_inputDescription :: Lens' UpdateInput (Maybe Text)

A brief description of the input.

updateInput_inputName :: Lens' UpdateInput Text

The name of the input you want to update.

updateInput_inputDefinition :: Lens' UpdateInput InputDefinition

The definition of the input.

updateInputResponse_inputConfiguration :: Lens' UpdateInputResponse (Maybe InputConfiguration)

Information about the configuration of the input.

updateInputResponse_httpStatus :: Lens' UpdateInputResponse Int

The response's http status code.

ListAlarmModelVersions

listAlarmModelVersions_nextToken :: Lens' ListAlarmModelVersions (Maybe Text)

The token that you can use to return the next set of results.

listAlarmModelVersions_maxResults :: Lens' ListAlarmModelVersions (Maybe Natural)

The maximum number of results to be returned per request.

listAlarmModelVersions_alarmModelName :: Lens' ListAlarmModelVersions Text

The name of the alarm model.

listAlarmModelVersionsResponse_alarmModelVersionSummaries :: Lens' ListAlarmModelVersionsResponse (Maybe [AlarmModelVersionSummary])

A list that summarizes each alarm model version.

listAlarmModelVersionsResponse_nextToken :: Lens' ListAlarmModelVersionsResponse (Maybe Text)

The token that you can use to return the next set of results, or null if there are no more results.

listAlarmModelVersionsResponse_httpStatus :: Lens' ListAlarmModelVersionsResponse Int

The response's http status code.

TagResource

tagResource_resourceArn :: Lens' TagResource Text

The ARN of the resource.

tagResource_tags :: Lens' TagResource [Tag]

The new or modified tags for the resource.

tagResourceResponse_httpStatus :: Lens' TagResourceResponse Int

The response's http status code.

DescribeDetectorModel

describeDetectorModel_detectorModelVersion :: Lens' DescribeDetectorModel (Maybe Text)

The version of the detector model.

describeDetectorModel_detectorModelName :: Lens' DescribeDetectorModel Text

The name of the detector model.

describeDetectorModelResponse_detectorModel :: Lens' DescribeDetectorModelResponse (Maybe DetectorModel)

Information about the detector model.

describeDetectorModelResponse_httpStatus :: Lens' DescribeDetectorModelResponse Int

The response's http status code.

DescribeInput

describeInput_inputName :: Lens' DescribeInput Text

The name of the input.

describeInputResponse_input :: Lens' DescribeInputResponse (Maybe Input)

Information about the input.

describeInputResponse_httpStatus :: Lens' DescribeInputResponse Int

The response's http status code.

UntagResource

untagResource_resourceArn :: Lens' UntagResource Text

The ARN of the resource.

untagResource_tagKeys :: Lens' UntagResource [Text]

A list of the keys of the tags to be removed from the resource.

untagResourceResponse_httpStatus :: Lens' UntagResourceResponse Int

The response's http status code.

ListInputRoutings

listInputRoutings_nextToken :: Lens' ListInputRoutings (Maybe Text)

The token that you can use to return the next set of results.

listInputRoutings_maxResults :: Lens' ListInputRoutings (Maybe Natural)

The maximum number of results to be returned per request.

listInputRoutings_inputIdentifier :: Lens' ListInputRoutings InputIdentifier

The identifer of the routed input.

listInputRoutingsResponse_routedResources :: Lens' ListInputRoutingsResponse (Maybe [RoutedResource])

Summary information about the routed resources.

listInputRoutingsResponse_nextToken :: Lens' ListInputRoutingsResponse (Maybe Text)

The token that you can use to return the next set of results, or null if there are no more results.

listInputRoutingsResponse_httpStatus :: Lens' ListInputRoutingsResponse Int

The response's http status code.

DescribeLoggingOptions

describeLoggingOptionsResponse_loggingOptions :: Lens' DescribeLoggingOptionsResponse (Maybe LoggingOptions)

The current settings of the AWS IoT Events logging options.

describeLoggingOptionsResponse_httpStatus :: Lens' DescribeLoggingOptionsResponse Int

The response's http status code.

Types

AcknowledgeFlow

acknowledgeFlow_enabled :: Lens' AcknowledgeFlow Bool

The value must be TRUE or FALSE. If TRUE, you receive a notification when the alarm state changes. You must choose to acknowledge the notification before the alarm state can return to NORMAL. If FALSE, you won't receive notifications. The alarm automatically changes to the NORMAL state when the input property value returns to the specified range.

Action

action_iotTopicPublish :: Lens' Action (Maybe IotTopicPublishAction)

Publishes an MQTT message with the given topic to the AWS IoT message broker.

action_dynamoDBv2 :: Lens' Action (Maybe DynamoDBv2Action)

Writes to the DynamoDB table that you created. The default action payload contains all attribute-value pairs that have the information about the detector model instance and the event that triggered the action. You can customize the payload. A separate column of the DynamoDB table receives one attribute-value pair in the payload that you specify. For more information, see Actions in AWS IoT Events Developer Guide.

action_resetTimer :: Lens' Action (Maybe ResetTimerAction)

Information needed to reset the timer.

action_setTimer :: Lens' Action (Maybe SetTimerAction)

Information needed to set the timer.

action_sns :: Lens' Action (Maybe SNSTopicPublishAction)

Sends an Amazon SNS message.

action_clearTimer :: Lens' Action (Maybe ClearTimerAction)

Information needed to clear the timer.

action_dynamoDB :: Lens' Action (Maybe DynamoDBAction)

Writes to the DynamoDB table that you created. The default action payload contains all attribute-value pairs that have the information about the detector model instance and the event that triggered the action. You can customize the payload. One column of the DynamoDB table receives all attribute-value pairs in the payload that you specify. For more information, see Actions in AWS IoT Events Developer Guide.

action_firehose :: Lens' Action (Maybe FirehoseAction)

Sends information about the detector model instance and the event that triggered the action to an Amazon Kinesis Data Firehose delivery stream.

action_iotSiteWise :: Lens' Action (Maybe IotSiteWiseAction)

Sends information about the detector model instance and the event that triggered the action to an asset property in AWS IoT SiteWise .

action_setVariable :: Lens' Action (Maybe SetVariableAction)

Sets a variable to a specified value.

action_lambda :: Lens' Action (Maybe LambdaAction)

Calls a Lambda function, passing in information about the detector model instance and the event that triggered the action.

action_iotEvents :: Lens' Action (Maybe IotEventsAction)

Sends AWS IoT Events input, which passes information about the detector model instance and the event that triggered the action.

action_sqs :: Lens' Action (Maybe SqsAction)

Sends information about the detector model instance and the event that triggered the action to an Amazon SQS queue.

AlarmAction

alarmAction_iotTopicPublish :: Lens' AlarmAction (Maybe IotTopicPublishAction)

Undocumented member.

alarmAction_dynamoDBv2 :: Lens' AlarmAction (Maybe DynamoDBv2Action)

Undocumented member.

alarmAction_sns :: Lens' AlarmAction (Maybe SNSTopicPublishAction)

Undocumented member.

alarmAction_dynamoDB :: Lens' AlarmAction (Maybe DynamoDBAction)

Undocumented member.

alarmAction_firehose :: Lens' AlarmAction (Maybe FirehoseAction)

Undocumented member.

alarmAction_iotSiteWise :: Lens' AlarmAction (Maybe IotSiteWiseAction)

Undocumented member.

alarmAction_lambda :: Lens' AlarmAction (Maybe LambdaAction)

Undocumented member.

alarmAction_iotEvents :: Lens' AlarmAction (Maybe IotEventsAction)

Undocumented member.

alarmAction_sqs :: Lens' AlarmAction (Maybe SqsAction)

Undocumented member.

AlarmCapabilities

alarmCapabilities_acknowledgeFlow :: Lens' AlarmCapabilities (Maybe AcknowledgeFlow)

Specifies whether to get notified for alarm state changes.

alarmCapabilities_initializationConfiguration :: Lens' AlarmCapabilities (Maybe InitializationConfiguration)

Specifies the default alarm state. The configuration applies to all alarms that were created based on this alarm model.

AlarmEventActions

alarmEventActions_alarmActions :: Lens' AlarmEventActions (Maybe [AlarmAction])

Specifies one or more supported actions to receive notifications when the alarm state changes.

AlarmModelSummary

alarmModelSummary_creationTime :: Lens' AlarmModelSummary (Maybe UTCTime)

The time the alarm model was created, in the Unix epoch format.

alarmModelSummary_alarmModelName :: Lens' AlarmModelSummary (Maybe Text)

The name of the alarm model.

alarmModelSummary_alarmModelDescription :: Lens' AlarmModelSummary (Maybe Text)

The description of the alarm model.

AlarmModelVersionSummary

alarmModelVersionSummary_creationTime :: Lens' AlarmModelVersionSummary (Maybe UTCTime)

The time the alarm model was created, in the Unix epoch format.

alarmModelVersionSummary_status :: Lens' AlarmModelVersionSummary (Maybe AlarmModelVersionStatus)

The status of the alarm model. The status can be one of the following values:

ACTIVE - The alarm model is active and it's ready to evaluate data.

ACTIVATING - AWS IoT Events is activating your alarm model. Activating an alarm model can take up to a few minutes.

INACTIVE - The alarm model is inactive, so it isn't ready to evaluate data. Check your alarm model information and update the alarm model.

FAILED - You couldn't create or update the alarm model. Check your alarm model information and try again.

alarmModelVersionSummary_alarmModelName :: Lens' AlarmModelVersionSummary (Maybe Text)

The name of the alarm model.

alarmModelVersionSummary_statusMessage :: Lens' AlarmModelVersionSummary (Maybe Text)

Contains information about the status of the alarm model version.

alarmModelVersionSummary_lastUpdateTime :: Lens' AlarmModelVersionSummary (Maybe UTCTime)

The time the alarm model was last updated, in the Unix epoch format.

alarmModelVersionSummary_alarmModelVersion :: Lens' AlarmModelVersionSummary (Maybe Text)

The version of the alarm model.

alarmModelVersionSummary_alarmModelArn :: Lens' AlarmModelVersionSummary (Maybe Text)

The ARN of the alarm model. For more information, see Amazon Resource Names (ARNs) in the AWS General Reference.

alarmModelVersionSummary_roleArn :: Lens' AlarmModelVersionSummary (Maybe Text)

The ARN of the IAM role that allows the alarm to perform actions and access AWS resources. For more information, see Amazon Resource Names (ARNs) in the AWS General Reference.

AlarmNotification

alarmNotification_notificationActions :: Lens' AlarmNotification (Maybe (NonEmpty NotificationAction))

Contains the notification settings of an alarm model. The settings apply to all alarms that were created based on this alarm model.

AlarmRule

alarmRule_simpleRule :: Lens' AlarmRule (Maybe SimpleRule)

A rule that compares an input property value to a threshold value with a comparison operator.

AnalysisResult

analysisResult_type :: Lens' AnalysisResult (Maybe Text)

The type of the analysis result. Analyses fall into the following types based on the validators used to generate the analysis result:

supported-actions - You must specify AWS IoT Events supported actions that work with other AWS services in a supported AWS Region.

service-limits - Resources or API operations can't exceed service quotas (also known as limits). Update your detector model or request a quota increase.

structure - The detector model must follow a structure that AWS IoT Events supports.

expression-syntax - Your expression must follow the required syntax.

data-type - Data types referenced in the detector model must be compatible.

referenced-data - You must define the data referenced in your detector model before you can use the data.

referenced-resource - Resources that the detector model uses must be available.

For more information, see Running detector model analyses in the AWS IoT Events Developer Guide.

analysisResult_message :: Lens' AnalysisResult (Maybe Text)

Contains additional information about the analysis result.

analysisResult_level :: Lens' AnalysisResult (Maybe AnalysisResultLevel)

The severity level of the analysis result. Based on the severity level, analysis results fall into three general categories:

INFO - An information result tells you about a significant field in your detector model. This type of result usually doesn't require immediate action.

WARNING - A warning result draws special attention to fields that might cause issues for your detector model. We recommend that you review warnings and take necessary actions before you use your detector model in production environments. Otherwise, the detector model might not work as expected.

ERROR - An error result notifies you about a problem found in your detector model. You must fix all errors before you can publish your detector model.

analysisResult_locations :: Lens' AnalysisResult (Maybe [AnalysisResultLocation])

Contains one or more locations that you can use to locate the fields in your detector model that the analysis result references.

AnalysisResultLocation

analysisResultLocation_path :: Lens' AnalysisResultLocation (Maybe Text)

A JsonPath expression that identifies the error field in your detector model.

AssetPropertyTimestamp

assetPropertyTimestamp_offsetInNanos :: Lens' AssetPropertyTimestamp (Maybe Text)

The nanosecond offset converted from timeInSeconds. The valid range is between 0-999999999.

assetPropertyTimestamp_timeInSeconds :: Lens' AssetPropertyTimestamp Text

The timestamp, in seconds, in the Unix epoch format. The valid range is between 1-31556889864403199.

AssetPropertyValue

assetPropertyValue_value :: Lens' AssetPropertyValue (Maybe AssetPropertyVariant)

The value to send to an asset property.

assetPropertyValue_quality :: Lens' AssetPropertyValue (Maybe Text)

The quality of the asset property value. The value must be 'GOOD', 'BAD', or 'UNCERTAIN'.

assetPropertyValue_timestamp :: Lens' AssetPropertyValue (Maybe AssetPropertyTimestamp)

The timestamp associated with the asset property value. The default is the current event time.

AssetPropertyVariant

assetPropertyVariant_integerValue :: Lens' AssetPropertyVariant (Maybe Text)

The asset property value is an integer. You must use an expression, and the evaluated result should be an integer.

assetPropertyVariant_doubleValue :: Lens' AssetPropertyVariant (Maybe Text)

The asset property value is a double. You must use an expression, and the evaluated result should be a double.

assetPropertyVariant_stringValue :: Lens' AssetPropertyVariant (Maybe Text)

The asset property value is a string. You must use an expression, and the evaluated result should be a string.

assetPropertyVariant_booleanValue :: Lens' AssetPropertyVariant (Maybe Text)

The asset property value is a Boolean value that must be 'TRUE' or 'FALSE'. You must use an expression, and the evaluated result should be a Boolean value.

Attribute

attribute_jsonPath :: Lens' Attribute Text

An expression that specifies an attribute-value pair in a JSON structure. Use this to specify an attribute from the JSON payload that is made available by the input. Inputs are derived from messages sent to AWS IoT Events (BatchPutMessage). Each such message contains a JSON payload. The attribute (and its paired value) specified here are available for use in the condition expressions used by detectors.

Syntax: <field-name>.<field-name>...

ClearTimerAction

clearTimerAction_timerName :: Lens' ClearTimerAction Text

The name of the timer to clear.

DetectorDebugOption

detectorDebugOption_keyValue :: Lens' DetectorDebugOption (Maybe Text)

The value of the input attribute key used to create the detector (the instance of the detector model).

detectorDebugOption_detectorModelName :: Lens' DetectorDebugOption Text

The name of the detector model.

DetectorModel

detectorModel_detectorModelConfiguration :: Lens' DetectorModel (Maybe DetectorModelConfiguration)

Information about how the detector is configured.

detectorModel_detectorModelDefinition :: Lens' DetectorModel (Maybe DetectorModelDefinition)

Information that defines how a detector operates.

DetectorModelConfiguration

detectorModelConfiguration_creationTime :: Lens' DetectorModelConfiguration (Maybe UTCTime)

The time the detector model was created.

detectorModelConfiguration_status :: Lens' DetectorModelConfiguration (Maybe DetectorModelVersionStatus)

The status of the detector model.

detectorModelConfiguration_detectorModelName :: Lens' DetectorModelConfiguration (Maybe Text)

The name of the detector model.

detectorModelConfiguration_detectorModelArn :: Lens' DetectorModelConfiguration (Maybe Text)

The ARN of the detector model.

detectorModelConfiguration_detectorModelDescription :: Lens' DetectorModelConfiguration (Maybe Text)

A brief description of the detector model.

detectorModelConfiguration_key :: Lens' DetectorModelConfiguration (Maybe Text)

The value used to identify a detector instance. When a device or system sends input, a new detector instance with a unique key value is created. AWS IoT Events can continue to route input to its corresponding detector instance based on this identifying information.

This parameter uses a JSON-path expression to select the attribute-value pair in the message payload that is used for identification. To route the message to the correct detector instance, the device must send a message payload that contains the same attribute-value.

detectorModelConfiguration_detectorModelVersion :: Lens' DetectorModelConfiguration (Maybe Text)

The version of the detector model.

detectorModelConfiguration_lastUpdateTime :: Lens' DetectorModelConfiguration (Maybe UTCTime)

The time the detector model was last updated.

detectorModelConfiguration_evaluationMethod :: Lens' DetectorModelConfiguration (Maybe EvaluationMethod)

Information about the order in which events are evaluated and how actions are executed.

detectorModelConfiguration_roleArn :: Lens' DetectorModelConfiguration (Maybe Text)

The ARN of the role that grants permission to AWS IoT Events to perform its operations.

DetectorModelDefinition

detectorModelDefinition_states :: Lens' DetectorModelDefinition (NonEmpty State)

Information about the states of the detector.

detectorModelDefinition_initialStateName :: Lens' DetectorModelDefinition Text

The state that is entered at the creation of each detector (instance).

DetectorModelSummary

detectorModelSummary_creationTime :: Lens' DetectorModelSummary (Maybe UTCTime)

The time the detector model was created.

detectorModelSummary_detectorModelName :: Lens' DetectorModelSummary (Maybe Text)

The name of the detector model.

detectorModelSummary_detectorModelDescription :: Lens' DetectorModelSummary (Maybe Text)

A brief description of the detector model.

DetectorModelVersionSummary

detectorModelVersionSummary_creationTime :: Lens' DetectorModelVersionSummary (Maybe UTCTime)

The time the detector model version was created.

detectorModelVersionSummary_status :: Lens' DetectorModelVersionSummary (Maybe DetectorModelVersionStatus)

The status of the detector model version.

detectorModelVersionSummary_detectorModelName :: Lens' DetectorModelVersionSummary (Maybe Text)

The name of the detector model.

detectorModelVersionSummary_detectorModelArn :: Lens' DetectorModelVersionSummary (Maybe Text)

The ARN of the detector model version.

detectorModelVersionSummary_detectorModelVersion :: Lens' DetectorModelVersionSummary (Maybe Text)

The ID of the detector model version.

detectorModelVersionSummary_lastUpdateTime :: Lens' DetectorModelVersionSummary (Maybe UTCTime)

The last time the detector model version was updated.

detectorModelVersionSummary_evaluationMethod :: Lens' DetectorModelVersionSummary (Maybe EvaluationMethod)

Information about the order in which events are evaluated and how actions are executed.

detectorModelVersionSummary_roleArn :: Lens' DetectorModelVersionSummary (Maybe Text)

The ARN of the role that grants the detector model permission to perform its tasks.

DynamoDBAction

dynamoDBAction_hashKeyType :: Lens' DynamoDBAction (Maybe Text)

The data type for the hash key (also called the partition key). You can specify the following values:

'STRING' - The hash key is a string.

'NUMBER' - The hash key is a number.

If you don't specify hashKeyType, the default value is 'STRING'.

dynamoDBAction_operation :: Lens' DynamoDBAction (Maybe Text)

The type of operation to perform. You can specify the following values:

'INSERT' - Insert data as a new item into the DynamoDB table. This item uses the specified hash key as a partition key. If you specified a range key, the item uses the range key as a sort key.

'UPDATE' - Update an existing item of the DynamoDB table with new data. This item's partition key must match the specified hash key. If you specified a range key, the range key must match the item's sort key.

'DELETE' - Delete an existing item of the DynamoDB table. This item's partition key must match the specified hash key. If you specified a range key, the range key must match the item's sort key.

If you don't specify this parameter, AWS IoT Events triggers the 'INSERT' operation.

dynamoDBAction_rangeKeyType :: Lens' DynamoDBAction (Maybe Text)

The data type for the range key (also called the sort key), You can specify the following values:

'STRING' - The range key is a string.

'NUMBER' - The range key is number.

If you don't specify rangeKeyField, the default value is 'STRING'.

dynamoDBAction_payload :: Lens' DynamoDBAction (Maybe Payload)

Undocumented member.

dynamoDBAction_payloadField :: Lens' DynamoDBAction (Maybe Text)

The name of the DynamoDB column that receives the action payload.

If you don't specify this parameter, the name of the DynamoDB column is payload.

dynamoDBAction_rangeKeyField :: Lens' DynamoDBAction (Maybe Text)

The name of the range key (also called the sort key). The rangeKeyField value must match the sort key of the target DynamoDB table.

dynamoDBAction_rangeKeyValue :: Lens' DynamoDBAction (Maybe Text)

The value of the range key (also called the sort key).

dynamoDBAction_hashKeyField :: Lens' DynamoDBAction Text

The name of the hash key (also called the partition key). The hashKeyField value must match the partition key of the target DynamoDB table.

dynamoDBAction_hashKeyValue :: Lens' DynamoDBAction Text

The value of the hash key (also called the partition key).

dynamoDBAction_tableName :: Lens' DynamoDBAction Text

The name of the DynamoDB table. The tableName value must match the table name of the target DynamoDB table.

DynamoDBv2Action

dynamoDBv2Action_payload :: Lens' DynamoDBv2Action (Maybe Payload)

Undocumented member.

dynamoDBv2Action_tableName :: Lens' DynamoDBv2Action Text

The name of the DynamoDB table.

EmailConfiguration

emailConfiguration_content :: Lens' EmailConfiguration (Maybe EmailContent)

Contains the subject and message of an email.

emailConfiguration_from :: Lens' EmailConfiguration Text

The email address that sends emails.

If you use the AWS IoT Events managed AWS Lambda function to manage your emails, you must verify the email address that sends emails in Amazon SES.

emailConfiguration_recipients :: Lens' EmailConfiguration EmailRecipients

Contains the information of one or more recipients who receive the emails.

You must add the users that receive emails to your AWS SSO store.

EmailContent

emailContent_subject :: Lens' EmailContent (Maybe Text)

The subject of the email.

emailContent_additionalMessage :: Lens' EmailContent (Maybe Text)

The message that you want to send. The message can be up to 200 characters.

EmailRecipients

emailRecipients_to :: Lens' EmailRecipients (Maybe (NonEmpty RecipientDetail))

Specifies one or more recipients who receive the email.

Event

event_actions :: Lens' Event (Maybe [Action])

The actions to be performed.

event_condition :: Lens' Event (Maybe Text)

Optional. The Boolean expression that, when TRUE, causes the actions to be performed. If not present, the actions are performed (=TRUE). If the expression result is not a Boolean value, the actions are not performed (=FALSE).

event_eventName :: Lens' Event Text

The name of the event.

FirehoseAction

firehoseAction_separator :: Lens' FirehoseAction (Maybe Text)

A character separator that is used to separate records written to the Kinesis Data Firehose delivery stream. Valid values are: '\n' (newline), '\t' (tab), '\r\n' (Windows newline), ',' (comma).

firehoseAction_payload :: Lens' FirehoseAction (Maybe Payload)

You can configure the action payload when you send a message to an Amazon Kinesis Data Firehose delivery stream.

firehoseAction_deliveryStreamName :: Lens' FirehoseAction Text

The name of the Kinesis Data Firehose delivery stream where the data is written.

InitializationConfiguration

initializationConfiguration_disabledOnInitialization :: Lens' InitializationConfiguration Bool

The value must be TRUE or FALSE. If FALSE, all alarm instances created based on the alarm model are activated. The default value is TRUE.

Input

input_inputDefinition :: Lens' Input (Maybe InputDefinition)

The definition of the input.

input_inputConfiguration :: Lens' Input (Maybe InputConfiguration)

Information about the configuration of an input.

InputConfiguration

inputConfiguration_inputDescription :: Lens' InputConfiguration (Maybe Text)

A brief description of the input.

inputConfiguration_inputName :: Lens' InputConfiguration Text

The name of the input.

inputConfiguration_inputArn :: Lens' InputConfiguration Text

The ARN of the input.

inputConfiguration_creationTime :: Lens' InputConfiguration UTCTime

The time the input was created.

inputConfiguration_lastUpdateTime :: Lens' InputConfiguration UTCTime

The last time the input was updated.

inputConfiguration_status :: Lens' InputConfiguration InputStatus

The status of the input.

InputDefinition

inputDefinition_attributes :: Lens' InputDefinition (NonEmpty Attribute)

The attributes from the JSON payload that are made available by the input. Inputs are derived from messages sent to the AWS IoT Events system using BatchPutMessage. Each such message contains a JSON payload, and those attributes (and their paired values) specified here are available for use in the condition expressions used by detectors that monitor this input.

InputIdentifier

inputIdentifier_iotSiteWiseInputIdentifier :: Lens' InputIdentifier (Maybe IotSiteWiseInputIdentifier)

The identifer of the input routed from AWS IoT SiteWise.

inputIdentifier_iotEventsInputIdentifier :: Lens' InputIdentifier (Maybe IotEventsInputIdentifier)

The identifier of the input routed to AWS IoT Events.

InputSummary

inputSummary_creationTime :: Lens' InputSummary (Maybe UTCTime)

The time the input was created.

inputSummary_status :: Lens' InputSummary (Maybe InputStatus)

The status of the input.

inputSummary_inputName :: Lens' InputSummary (Maybe Text)

The name of the input.

inputSummary_inputDescription :: Lens' InputSummary (Maybe Text)

A brief description of the input.

inputSummary_inputArn :: Lens' InputSummary (Maybe Text)

The ARN of the input.

inputSummary_lastUpdateTime :: Lens' InputSummary (Maybe UTCTime)

The last time the input was updated.

IotEventsAction

iotEventsAction_payload :: Lens' IotEventsAction (Maybe Payload)

You can configure the action payload when you send a message to an AWS IoT Events input.

iotEventsAction_inputName :: Lens' IotEventsAction Text

The name of the AWS IoT Events input where the data is sent.

IotEventsInputIdentifier

iotEventsInputIdentifier_inputName :: Lens' IotEventsInputIdentifier Text

The name of the input routed to AWS IoT Events.

IotSiteWiseAction

iotSiteWiseAction_entryId :: Lens' IotSiteWiseAction (Maybe Text)

A unique identifier for this entry. You can use the entry ID to track which data entry causes an error in case of failure. The default is a new unique identifier.

iotSiteWiseAction_propertyAlias :: Lens' IotSiteWiseAction (Maybe Text)

The alias of the asset property.

iotSiteWiseAction_propertyValue :: Lens' IotSiteWiseAction (Maybe AssetPropertyValue)

The value to send to the asset property. This value contains timestamp, quality, and value (TQV) information.

iotSiteWiseAction_propertyId :: Lens' IotSiteWiseAction (Maybe Text)

The ID of the asset property.

iotSiteWiseAction_assetId :: Lens' IotSiteWiseAction (Maybe Text)

The ID of the asset that has the specified property.

IotSiteWiseAssetModelPropertyIdentifier

iotSiteWiseAssetModelPropertyIdentifier_assetModelId :: Lens' IotSiteWiseAssetModelPropertyIdentifier Text

The ID of the AWS IoT SiteWise asset model.

iotSiteWiseAssetModelPropertyIdentifier_propertyId :: Lens' IotSiteWiseAssetModelPropertyIdentifier Text

The ID of the AWS IoT SiteWise asset property.

IotSiteWiseInputIdentifier

iotSiteWiseInputIdentifier_iotSiteWiseAssetModelPropertyIdentifier :: Lens' IotSiteWiseInputIdentifier (Maybe IotSiteWiseAssetModelPropertyIdentifier)

The identifier of the AWS IoT SiteWise asset model property.

IotTopicPublishAction

iotTopicPublishAction_payload :: Lens' IotTopicPublishAction (Maybe Payload)

You can configure the action payload when you publish a message to an AWS IoT Core topic.

iotTopicPublishAction_mqttTopic :: Lens' IotTopicPublishAction Text

The MQTT topic of the message. You can use a string expression that includes variables ($variable.<variable-name>) and input values ($input.<input-name>.<path-to-datum>) as the topic string.

LambdaAction

lambdaAction_payload :: Lens' LambdaAction (Maybe Payload)

You can configure the action payload when you send a message to a Lambda function.

lambdaAction_functionArn :: Lens' LambdaAction Text

The ARN of the Lambda function that is executed.

LoggingOptions

loggingOptions_detectorDebugOptions :: Lens' LoggingOptions (Maybe (NonEmpty DetectorDebugOption))

Information that identifies those detector models and their detectors (instances) for which the logging level is given.

loggingOptions_roleArn :: Lens' LoggingOptions Text

The ARN of the role that grants permission to AWS IoT Events to perform logging.

loggingOptions_level :: Lens' LoggingOptions LoggingLevel

The logging level.

loggingOptions_enabled :: Lens' LoggingOptions Bool

If TRUE, logging is enabled for AWS IoT Events.

NotificationAction

notificationAction_emailConfigurations :: Lens' NotificationAction (Maybe (NonEmpty EmailConfiguration))

Contains the configuration information of email notifications.

notificationAction_smsConfigurations :: Lens' NotificationAction (Maybe (NonEmpty SMSConfiguration))

Contains the configuration information of SMS notifications.

notificationAction_action :: Lens' NotificationAction NotificationTargetActions

Specifies an AWS Lambda function to manage alarm notifications. You can create one or use the AWS Lambda function provided by AWS IoT Events.

NotificationTargetActions

notificationTargetActions_lambdaAction :: Lens' NotificationTargetActions (Maybe LambdaAction)

Undocumented member.

OnEnterLifecycle

onEnterLifecycle_events :: Lens' OnEnterLifecycle (Maybe [Event])

Specifies the actions that are performed when the state is entered and the condition is TRUE.

OnExitLifecycle

onExitLifecycle_events :: Lens' OnExitLifecycle (Maybe [Event])

Specifies the actions that are performed when the state is exited and the condition is TRUE.

OnInputLifecycle

onInputLifecycle_events :: Lens' OnInputLifecycle (Maybe [Event])

Specifies the actions performed when the condition evaluates to TRUE.

onInputLifecycle_transitionEvents :: Lens' OnInputLifecycle (Maybe [TransitionEvent])

Specifies the actions performed, and the next state entered, when a condition evaluates to TRUE.

Payload

payload_contentExpression :: Lens' Payload Text

The content of the payload. You can use a string expression that includes quoted strings ('<string>'), variables ($variable.<variable-name>), input values ($input.<input-name>.<path-to-datum>), string concatenations, and quoted strings that contain ${} as the content. The recommended maximum size of a content expression is 1 KB.

payload_type :: Lens' Payload PayloadType

The value of the payload type can be either STRING or JSON.

RecipientDetail

recipientDetail_ssoIdentity :: Lens' RecipientDetail (Maybe SSOIdentity)

The AWS Single Sign-On (AWS SSO) authentication information.

ResetTimerAction

resetTimerAction_timerName :: Lens' ResetTimerAction Text

The name of the timer to reset.

RoutedResource

routedResource_arn :: Lens' RoutedResource (Maybe Text)

The ARN of the routed resource. For more information, see Amazon Resource Names (ARNs) in the AWS General Reference.

routedResource_name :: Lens' RoutedResource (Maybe Text)

The name of the routed resource.

SMSConfiguration

sMSConfiguration_additionalMessage :: Lens' SMSConfiguration (Maybe Text)

The message that you want to send. The message can be up to 200 characters.

sMSConfiguration_senderId :: Lens' SMSConfiguration (Maybe Text)

The sender ID.

sMSConfiguration_recipients :: Lens' SMSConfiguration (NonEmpty RecipientDetail)

Specifies one or more recipients who receive the message.

You must add the users that receive SMS messages to your AWS SSO store.

SNSTopicPublishAction

sNSTopicPublishAction_payload :: Lens' SNSTopicPublishAction (Maybe Payload)

You can configure the action payload when you send a message as an Amazon SNS push notification.

sNSTopicPublishAction_targetArn :: Lens' SNSTopicPublishAction Text

The ARN of the Amazon SNS target where the message is sent.

SSOIdentity

sSOIdentity_userId :: Lens' SSOIdentity (Maybe Text)

The user ID.

sSOIdentity_identityStoreId :: Lens' SSOIdentity Text

The ID of the AWS SSO identity store.

SetTimerAction

setTimerAction_durationExpression :: Lens' SetTimerAction (Maybe Text)

The duration of the timer, in seconds. You can use a string expression that includes numbers, variables ($variable.<variable-name>), and input values ($input.<input-name>.<path-to-datum>) as the duration. The range of the duration is 1-31622400 seconds. To ensure accuracy, the minimum duration is 60 seconds. The evaluated result of the duration is rounded down to the nearest whole number.

setTimerAction_seconds :: Lens' SetTimerAction (Maybe Natural)

The number of seconds until the timer expires. The minimum value is 60 seconds to ensure accuracy. The maximum value is 31622400 seconds.

setTimerAction_timerName :: Lens' SetTimerAction Text

The name of the timer.

SetVariableAction

setVariableAction_variableName :: Lens' SetVariableAction Text

The name of the variable.

setVariableAction_value :: Lens' SetVariableAction Text

The new value of the variable.

SimpleRule

simpleRule_inputProperty :: Lens' SimpleRule Text

The value on the left side of the comparison operator. You can specify an AWS IoT Events input attribute as an input property.

simpleRule_comparisonOperator :: Lens' SimpleRule ComparisonOperator

The comparison operator.

simpleRule_threshold :: Lens' SimpleRule Text

The value on the right side of the comparison operator. You can enter a number or specify an AWS IoT Events input attribute.

SqsAction

sqsAction_payload :: Lens' SqsAction (Maybe Payload)

You can configure the action payload when you send a message to an Amazon SQS queue.

sqsAction_useBase64 :: Lens' SqsAction (Maybe Bool)

Set this to TRUE if you want the data to be base-64 encoded before it is written to the queue. Otherwise, set this to FALSE.

sqsAction_queueUrl :: Lens' SqsAction Text

The URL of the SQS queue where the data is written.

State

state_onEnter :: Lens' State (Maybe OnEnterLifecycle)

When entering this state, perform these actions if the condition is TRUE.

state_onInput :: Lens' State (Maybe OnInputLifecycle)

When an input is received and the condition is TRUE, perform the specified actions.

state_onExit :: Lens' State (Maybe OnExitLifecycle)

When exiting this state, perform these actions if the specified condition is TRUE.

state_stateName :: Lens' State Text

The name of the state.

Tag

tag_key :: Lens' Tag Text

The tag's key.

tag_value :: Lens' Tag Text

The tag's value.

TransitionEvent

transitionEvent_actions :: Lens' TransitionEvent (Maybe [Action])

The actions to be performed.

transitionEvent_eventName :: Lens' TransitionEvent Text

The name of the transition event.

transitionEvent_condition :: Lens' TransitionEvent Text

Required. A Boolean expression that when TRUE causes the actions to be performed and the nextState to be entered.

transitionEvent_nextState :: Lens' TransitionEvent Text

The next state to enter.