Amazonka.IoTEvents.Types.AlarmModelVersionSummary

Description

Documentation

data AlarmModelVersionSummary

Contains a summary of an alarm model version.

See: newAlarmModelVersionSummary smart constructor.

Constructors

AlarmModelVersionSummary'

Fields

creationTime :: Maybe POSIX

The time the alarm model was created, in the Unix epoch format.

status :: Maybe AlarmModelVersionStatus

The status of the alarm model. The status can be one of the following values:

ACTIVE - The alarm model is active and it's ready to evaluate data.

ACTIVATING - AWS IoT Events is activating your alarm model. Activating an alarm model can take up to a few minutes.

INACTIVE - The alarm model is inactive, so it isn't ready to evaluate data. Check your alarm model information and update the alarm model.

FAILED - You couldn't create or update the alarm model. Check your alarm model information and try again.

alarmModelName :: Maybe Text

The name of the alarm model.

statusMessage :: Maybe Text

Contains information about the status of the alarm model version.

lastUpdateTime :: Maybe POSIX

The time the alarm model was last updated, in the Unix epoch format.

alarmModelVersion :: Maybe Text

The version of the alarm model.

alarmModelArn :: Maybe Text

The ARN of the alarm model. For more information, see Amazon Resource Names (ARNs) in the AWS General Reference.

roleArn :: Maybe Text

The ARN of the IAM role that allows the alarm to perform actions and access AWS resources. For more information, see Amazon Resource Names (ARNs) in the AWS General Reference.

newAlarmModelVersionSummary :: AlarmModelVersionSummary

Create a value of AlarmModelVersionSummary with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:creationTime:AlarmModelVersionSummary', alarmModelVersionSummary_creationTime - The time the alarm model was created, in the Unix epoch format.

$sel:status:AlarmModelVersionSummary', alarmModelVersionSummary_status - The status of the alarm model. The status can be one of the following values:

ACTIVE - The alarm model is active and it's ready to evaluate data.

ACTIVATING - AWS IoT Events is activating your alarm model. Activating an alarm model can take up to a few minutes.

INACTIVE - The alarm model is inactive, so it isn't ready to evaluate data. Check your alarm model information and update the alarm model.

FAILED - You couldn't create or update the alarm model. Check your alarm model information and try again.

$sel:alarmModelName:AlarmModelVersionSummary', alarmModelVersionSummary_alarmModelName - The name of the alarm model.

$sel:statusMessage:AlarmModelVersionSummary', alarmModelVersionSummary_statusMessage - Contains information about the status of the alarm model version.

$sel:lastUpdateTime:AlarmModelVersionSummary', alarmModelVersionSummary_lastUpdateTime - The time the alarm model was last updated, in the Unix epoch format.

$sel:alarmModelVersion:AlarmModelVersionSummary', alarmModelVersionSummary_alarmModelVersion - The version of the alarm model.

$sel:alarmModelArn:AlarmModelVersionSummary', alarmModelVersionSummary_alarmModelArn - The ARN of the alarm model. For more information, see Amazon Resource Names (ARNs) in the AWS General Reference.

$sel:roleArn:AlarmModelVersionSummary', alarmModelVersionSummary_roleArn - The ARN of the IAM role that allows the alarm to perform actions and access AWS resources. For more information, see Amazon Resource Names (ARNs) in the AWS General Reference.

alarmModelVersionSummary_creationTime :: Lens' AlarmModelVersionSummary (Maybe UTCTime)

The time the alarm model was created, in the Unix epoch format.

alarmModelVersionSummary_status :: Lens' AlarmModelVersionSummary (Maybe AlarmModelVersionStatus)

The status of the alarm model. The status can be one of the following values:

ACTIVE - The alarm model is active and it's ready to evaluate data.

ACTIVATING - AWS IoT Events is activating your alarm model. Activating an alarm model can take up to a few minutes.

INACTIVE - The alarm model is inactive, so it isn't ready to evaluate data. Check your alarm model information and update the alarm model.

FAILED - You couldn't create or update the alarm model. Check your alarm model information and try again.

alarmModelVersionSummary_alarmModelName :: Lens' AlarmModelVersionSummary (Maybe Text)

The name of the alarm model.

alarmModelVersionSummary_statusMessage :: Lens' AlarmModelVersionSummary (Maybe Text)

Contains information about the status of the alarm model version.

alarmModelVersionSummary_lastUpdateTime :: Lens' AlarmModelVersionSummary (Maybe UTCTime)

The time the alarm model was last updated, in the Unix epoch format.

alarmModelVersionSummary_alarmModelVersion :: Lens' AlarmModelVersionSummary (Maybe Text)

The version of the alarm model.

alarmModelVersionSummary_alarmModelArn :: Lens' AlarmModelVersionSummary (Maybe Text)

The ARN of the alarm model. For more information, see Amazon Resource Names (ARNs) in the AWS General Reference.

alarmModelVersionSummary_roleArn :: Lens' AlarmModelVersionSummary (Maybe Text)

The ARN of the IAM role that allows the alarm to perform actions and access AWS resources. For more information, see Amazon Resource Names (ARNs) in the AWS General Reference.