Amazonka.IoTEvents.Types.DynamoDBAction

Description

Documentation

data DynamoDBAction

Defines an action to write to the Amazon DynamoDB table that you created. The standard action payload contains all the information about the detector model instance and the event that triggered the action. You can customize the payload. One column of the DynamoDB table receives all attribute-value pairs in the payload that you specify.

You must use expressions for all parameters in DynamoDBAction. The expressions accept literals, operators, functions, references, and substitution templates.

Examples

For literal values, the expressions must contain single quotes. For example, the value for the hashKeyType parameter can be 'STRING'.

For references, you must specify either variables or input values. For example, the value for the hashKeyField parameter can be $input.GreenhouseInput.name.

For a substitution template, you must use ${}, and the template must be in single quotes. A substitution template can also contain a combination of literals, operators, functions, references, and substitution templates.

In the following example, the value for the hashKeyValue parameter uses a substitution template.

'${$input.GreenhouseInput.temperature * 6 / 5 + 32} in Fahrenheit'

For a string concatenation, you must use +. A string concatenation can also contain a combination of literals, operators, functions, references, and substitution templates.

In the following example, the value for the tableName parameter uses a string concatenation.

'GreenhouseTemperatureTable ' + $input.GreenhouseInput.date

For more information, see Expressions in the AWS IoT Events Developer Guide.

If the defined payload type is a string, DynamoDBAction writes non-JSON data to the DynamoDB table as binary data. The DynamoDB console displays the data as Base64-encoded text. The value for the payloadField parameter is <payload-field>_raw.

See: newDynamoDBAction smart constructor.

Constructors

DynamoDBAction'

Fields

hashKeyType :: Maybe Text

The data type for the hash key (also called the partition key). You can specify the following values:

'STRING' - The hash key is a string.

'NUMBER' - The hash key is a number.

If you don't specify hashKeyType, the default value is 'STRING'.

operation :: Maybe Text

The type of operation to perform. You can specify the following values:

'INSERT' - Insert data as a new item into the DynamoDB table. This item uses the specified hash key as a partition key. If you specified a range key, the item uses the range key as a sort key.

'UPDATE' - Update an existing item of the DynamoDB table with new data. This item's partition key must match the specified hash key. If you specified a range key, the range key must match the item's sort key.

'DELETE' - Delete an existing item of the DynamoDB table. This item's partition key must match the specified hash key. If you specified a range key, the range key must match the item's sort key.

If you don't specify this parameter, AWS IoT Events triggers the 'INSERT' operation.

rangeKeyType :: Maybe Text

The data type for the range key (also called the sort key), You can specify the following values:

'STRING' - The range key is a string.

'NUMBER' - The range key is number.

If you don't specify rangeKeyField, the default value is 'STRING'.

payload :: Maybe Payload

payloadField :: Maybe Text

The name of the DynamoDB column that receives the action payload.

If you don't specify this parameter, the name of the DynamoDB column is payload.

rangeKeyField :: Maybe Text

The name of the range key (also called the sort key). The rangeKeyField value must match the sort key of the target DynamoDB table.

rangeKeyValue :: Maybe Text

The value of the range key (also called the sort key).

hashKeyField :: Text

The name of the hash key (also called the partition key). The hashKeyField value must match the partition key of the target DynamoDB table.

hashKeyValue :: Text

The value of the hash key (also called the partition key).

tableName :: Text

The name of the DynamoDB table. The tableName value must match the table name of the target DynamoDB table.

newDynamoDBAction

Arguments

:: Text
$sel:hashKeyField:DynamoDBAction'

-> Text
$sel:hashKeyValue:DynamoDBAction'

-> Text
$sel:tableName:DynamoDBAction'

-> DynamoDBAction

Create a value of DynamoDBAction with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:hashKeyType:DynamoDBAction', dynamoDBAction_hashKeyType - The data type for the hash key (also called the partition key). You can specify the following values:

'STRING' - The hash key is a string.

'NUMBER' - The hash key is a number.

If you don't specify hashKeyType, the default value is 'STRING'.

$sel:operation:DynamoDBAction', dynamoDBAction_operation - The type of operation to perform. You can specify the following values:

'INSERT' - Insert data as a new item into the DynamoDB table. This item uses the specified hash key as a partition key. If you specified a range key, the item uses the range key as a sort key.

'UPDATE' - Update an existing item of the DynamoDB table with new data. This item's partition key must match the specified hash key. If you specified a range key, the range key must match the item's sort key.

'DELETE' - Delete an existing item of the DynamoDB table. This item's partition key must match the specified hash key. If you specified a range key, the range key must match the item's sort key.

If you don't specify this parameter, AWS IoT Events triggers the 'INSERT' operation.

$sel:rangeKeyType:DynamoDBAction', dynamoDBAction_rangeKeyType - The data type for the range key (also called the sort key), You can specify the following values:

'STRING' - The range key is a string.

'NUMBER' - The range key is number.

If you don't specify rangeKeyField, the default value is 'STRING'.

$sel:payload:DynamoDBAction', dynamoDBAction_payload - Undocumented member.

$sel:payloadField:DynamoDBAction', dynamoDBAction_payloadField - The name of the DynamoDB column that receives the action payload.

If you don't specify this parameter, the name of the DynamoDB column is payload.

$sel:rangeKeyField:DynamoDBAction', dynamoDBAction_rangeKeyField - The name of the range key (also called the sort key). The rangeKeyField value must match the sort key of the target DynamoDB table.

$sel:rangeKeyValue:DynamoDBAction', dynamoDBAction_rangeKeyValue - The value of the range key (also called the sort key).

$sel:hashKeyField:DynamoDBAction', dynamoDBAction_hashKeyField - The name of the hash key (also called the partition key). The hashKeyField value must match the partition key of the target DynamoDB table.

$sel:hashKeyValue:DynamoDBAction', dynamoDBAction_hashKeyValue - The value of the hash key (also called the partition key).

$sel:tableName:DynamoDBAction', dynamoDBAction_tableName - The name of the DynamoDB table. The tableName value must match the table name of the target DynamoDB table.

dynamoDBAction_hashKeyType :: Lens' DynamoDBAction (Maybe Text)

The data type for the hash key (also called the partition key). You can specify the following values:

'STRING' - The hash key is a string.

'NUMBER' - The hash key is a number.

If you don't specify hashKeyType, the default value is 'STRING'.

dynamoDBAction_operation :: Lens' DynamoDBAction (Maybe Text)

The type of operation to perform. You can specify the following values:

'INSERT' - Insert data as a new item into the DynamoDB table. This item uses the specified hash key as a partition key. If you specified a range key, the item uses the range key as a sort key.

'UPDATE' - Update an existing item of the DynamoDB table with new data. This item's partition key must match the specified hash key. If you specified a range key, the range key must match the item's sort key.

'DELETE' - Delete an existing item of the DynamoDB table. This item's partition key must match the specified hash key. If you specified a range key, the range key must match the item's sort key.

If you don't specify this parameter, AWS IoT Events triggers the 'INSERT' operation.

dynamoDBAction_rangeKeyType :: Lens' DynamoDBAction (Maybe Text)

The data type for the range key (also called the sort key), You can specify the following values:

'STRING' - The range key is a string.

'NUMBER' - The range key is number.

If you don't specify rangeKeyField, the default value is 'STRING'.

dynamoDBAction_payload :: Lens' DynamoDBAction (Maybe Payload)

Undocumented member.

dynamoDBAction_payloadField :: Lens' DynamoDBAction (Maybe Text)

The name of the DynamoDB column that receives the action payload.

If you don't specify this parameter, the name of the DynamoDB column is payload.

dynamoDBAction_rangeKeyField :: Lens' DynamoDBAction (Maybe Text)

The name of the range key (also called the sort key). The rangeKeyField value must match the sort key of the target DynamoDB table.

dynamoDBAction_rangeKeyValue :: Lens' DynamoDBAction (Maybe Text)

The value of the range key (also called the sort key).

dynamoDBAction_hashKeyField :: Lens' DynamoDBAction Text

The name of the hash key (also called the partition key). The hashKeyField value must match the partition key of the target DynamoDB table.

dynamoDBAction_hashKeyValue :: Lens' DynamoDBAction Text

The value of the hash key (also called the partition key).

dynamoDBAction_tableName :: Lens' DynamoDBAction Text

The name of the DynamoDB table. The tableName value must match the table name of the target DynamoDB table.