Amazonka.IoTEvents.Types.Event

Description

Documentation

data Event

Specifies the actions to be performed when the condition evaluates to TRUE.

See: newEvent smart constructor.

Constructors

Event'

Fields

actions :: Maybe [Action]

The actions to be performed.

condition :: Maybe Text

Optional. The Boolean expression that, when TRUE, causes the actions to be performed. If not present, the actions are performed (=TRUE). If the expression result is not a Boolean value, the actions are not performed (=FALSE).

eventName :: Text

The name of the event.

newEvent

Arguments

:: Text
$sel:eventName:Event'

-> Event

Create a value of Event with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:actions:Event', event_actions - The actions to be performed.

$sel:condition:Event', event_condition - Optional. The Boolean expression that, when TRUE, causes the actions to be performed. If not present, the actions are performed (=TRUE). If the expression result is not a Boolean value, the actions are not performed (=FALSE).

$sel:eventName:Event', event_eventName - The name of the event.

event_actions :: Lens' Event (Maybe [Action])

The actions to be performed.

event_condition :: Lens' Event (Maybe Text)

Optional. The Boolean expression that, when TRUE, causes the actions to be performed. If not present, the actions are performed (=TRUE). If the expression result is not a Boolean value, the actions are not performed (=FALSE).

event_eventName :: Lens' Event Text

The name of the event.