Amazonka.IoTEvents.Types.IotSiteWiseAction

Description

Documentation

data IotSiteWiseAction

Sends information about the detector model instance and the event that triggered the action to a specified asset property in AWS IoT SiteWise.

You must use expressions for all parameters in IotSiteWiseAction. The expressions accept literals, operators, functions, references, and substitutions templates.

Examples

For literal values, the expressions must contain single quotes. For example, the value for the propertyAlias parameter can be '/company/windfarm/3/turbine/7/temperature'.

For references, you must specify either variables or input values. For example, the value for the assetId parameter can be $input.TurbineInput.assetId1.

For a substitution template, you must use ${}, and the template must be in single quotes. A substitution template can also contain a combination of literals, operators, functions, references, and substitution templates.

In the following example, the value for the propertyAlias parameter uses a substitution template.

'company/windfarm/${$input.TemperatureInput.sensorData.windfarmID}/turbine/ ${$input.TemperatureInput.sensorData.turbineID}/temperature'

You must specify either propertyAlias or both assetId and propertyId to identify the target asset property in AWS IoT SiteWise.

For more information, see Expressions in the AWS IoT Events Developer Guide.

See: newIotSiteWiseAction smart constructor.

Constructors

IotSiteWiseAction'

Fields

entryId :: Maybe Text

A unique identifier for this entry. You can use the entry ID to track which data entry causes an error in case of failure. The default is a new unique identifier.

propertyAlias :: Maybe Text

The alias of the asset property.

propertyValue :: Maybe AssetPropertyValue

The value to send to the asset property. This value contains timestamp, quality, and value (TQV) information.

propertyId :: Maybe Text

The ID of the asset property.

assetId :: Maybe Text

The ID of the asset that has the specified property.

newIotSiteWiseAction :: IotSiteWiseAction

Create a value of IotSiteWiseAction with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:entryId:IotSiteWiseAction', iotSiteWiseAction_entryId - A unique identifier for this entry. You can use the entry ID to track which data entry causes an error in case of failure. The default is a new unique identifier.

$sel:propertyAlias:IotSiteWiseAction', iotSiteWiseAction_propertyAlias - The alias of the asset property.

$sel:propertyValue:IotSiteWiseAction', iotSiteWiseAction_propertyValue - The value to send to the asset property. This value contains timestamp, quality, and value (TQV) information.

$sel:propertyId:IotSiteWiseAction', iotSiteWiseAction_propertyId - The ID of the asset property.

$sel:assetId:IotSiteWiseAction', iotSiteWiseAction_assetId - The ID of the asset that has the specified property.

iotSiteWiseAction_entryId :: Lens' IotSiteWiseAction (Maybe Text)

A unique identifier for this entry. You can use the entry ID to track which data entry causes an error in case of failure. The default is a new unique identifier.

iotSiteWiseAction_propertyAlias :: Lens' IotSiteWiseAction (Maybe Text)

The alias of the asset property.

iotSiteWiseAction_propertyValue :: Lens' IotSiteWiseAction (Maybe AssetPropertyValue)

The value to send to the asset property. This value contains timestamp, quality, and value (TQV) information.

iotSiteWiseAction_propertyId :: Lens' IotSiteWiseAction (Maybe Text)

The ID of the asset property.

iotSiteWiseAction_assetId :: Lens' IotSiteWiseAction (Maybe Text)

The ID of the asset that has the specified property.