Amazonka.IoTEvents.UpdateAlarmModel

Description

Updates an alarm model. Any alarms that were created based on the previous version are deleted and then created again as new data arrives.

Creating a Request

data UpdateAlarmModel

See: newUpdateAlarmModel smart constructor.

Constructors

UpdateAlarmModel'

Fields

severity :: Maybe Natural

A non-negative integer that reflects the severity level of the alarm.

alarmNotification :: Maybe AlarmNotification

Contains information about one or more notification actions.

alarmModelDescription :: Maybe Text

The description of the alarm model.

alarmEventActions :: Maybe AlarmEventActions

Contains information about one or more alarm actions.

alarmCapabilities :: Maybe AlarmCapabilities

Contains the configuration information of alarm state changes.

alarmModelName :: Text

The name of the alarm model.

roleArn :: Text

The ARN of the IAM role that allows the alarm to perform actions and access AWS resources. For more information, see Amazon Resource Names (ARNs) in the AWS General Reference.

alarmRule :: AlarmRule

Defines when your alarm is invoked.

newUpdateAlarmModel

Arguments

:: Text
$sel:alarmModelName:UpdateAlarmModel'

-> Text
$sel:roleArn:UpdateAlarmModel'

-> AlarmRule
$sel:alarmRule:UpdateAlarmModel'

-> UpdateAlarmModel

Create a value of UpdateAlarmModel with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:severity:UpdateAlarmModel', updateAlarmModel_severity - A non-negative integer that reflects the severity level of the alarm.

$sel:alarmNotification:UpdateAlarmModel', updateAlarmModel_alarmNotification - Contains information about one or more notification actions.

$sel:alarmModelDescription:UpdateAlarmModel', updateAlarmModel_alarmModelDescription - The description of the alarm model.

$sel:alarmEventActions:UpdateAlarmModel', updateAlarmModel_alarmEventActions - Contains information about one or more alarm actions.

$sel:alarmCapabilities:UpdateAlarmModel', updateAlarmModel_alarmCapabilities - Contains the configuration information of alarm state changes.

$sel:alarmModelName:UpdateAlarmModel', updateAlarmModel_alarmModelName - The name of the alarm model.

$sel:roleArn:UpdateAlarmModel', updateAlarmModel_roleArn - The ARN of the IAM role that allows the alarm to perform actions and access AWS resources. For more information, see Amazon Resource Names (ARNs) in the AWS General Reference.

$sel:alarmRule:UpdateAlarmModel', updateAlarmModel_alarmRule - Defines when your alarm is invoked.

Request Lenses

updateAlarmModel_severity :: Lens' UpdateAlarmModel (Maybe Natural)

A non-negative integer that reflects the severity level of the alarm.

updateAlarmModel_alarmNotification :: Lens' UpdateAlarmModel (Maybe AlarmNotification)

Contains information about one or more notification actions.

updateAlarmModel_alarmModelDescription :: Lens' UpdateAlarmModel (Maybe Text)

The description of the alarm model.

updateAlarmModel_alarmEventActions :: Lens' UpdateAlarmModel (Maybe AlarmEventActions)

Contains information about one or more alarm actions.

updateAlarmModel_alarmCapabilities :: Lens' UpdateAlarmModel (Maybe AlarmCapabilities)

Contains the configuration information of alarm state changes.

updateAlarmModel_alarmModelName :: Lens' UpdateAlarmModel Text

The name of the alarm model.

updateAlarmModel_roleArn :: Lens' UpdateAlarmModel Text

The ARN of the IAM role that allows the alarm to perform actions and access AWS resources. For more information, see Amazon Resource Names (ARNs) in the AWS General Reference.

updateAlarmModel_alarmRule :: Lens' UpdateAlarmModel AlarmRule

Defines when your alarm is invoked.

Destructuring the Response

data UpdateAlarmModelResponse

See: newUpdateAlarmModelResponse smart constructor.

Constructors

UpdateAlarmModelResponse'

Fields

creationTime :: Maybe POSIX

The time the alarm model was created, in the Unix epoch format.

status :: Maybe AlarmModelVersionStatus

The status of the alarm model. The status can be one of the following values:

ACTIVE - The alarm model is active and it's ready to evaluate data.

ACTIVATING - AWS IoT Events is activating your alarm model. Activating an alarm model can take up to a few minutes.

INACTIVE - The alarm model is inactive, so it isn't ready to evaluate data. Check your alarm model information and update the alarm model.

FAILED - You couldn't create or update the alarm model. Check your alarm model information and try again.

lastUpdateTime :: Maybe POSIX

The time the alarm model was last updated, in the Unix epoch format.

alarmModelVersion :: Maybe Text

The version of the alarm model.

alarmModelArn :: Maybe Text

The ARN of the alarm model. For more information, see Amazon Resource Names (ARNs) in the AWS General Reference.

httpStatus :: Int

The response's http status code.

newUpdateAlarmModelResponse

Arguments

:: Int
$sel:httpStatus:UpdateAlarmModelResponse'

-> UpdateAlarmModelResponse

Create a value of UpdateAlarmModelResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:creationTime:UpdateAlarmModelResponse', updateAlarmModelResponse_creationTime - The time the alarm model was created, in the Unix epoch format.

$sel:status:UpdateAlarmModelResponse', updateAlarmModelResponse_status - The status of the alarm model. The status can be one of the following values:

ACTIVE - The alarm model is active and it's ready to evaluate data.

ACTIVATING - AWS IoT Events is activating your alarm model. Activating an alarm model can take up to a few minutes.

INACTIVE - The alarm model is inactive, so it isn't ready to evaluate data. Check your alarm model information and update the alarm model.

FAILED - You couldn't create or update the alarm model. Check your alarm model information and try again.

$sel:lastUpdateTime:UpdateAlarmModelResponse', updateAlarmModelResponse_lastUpdateTime - The time the alarm model was last updated, in the Unix epoch format.

$sel:alarmModelVersion:UpdateAlarmModelResponse', updateAlarmModelResponse_alarmModelVersion - The version of the alarm model.

$sel:alarmModelArn:UpdateAlarmModelResponse', updateAlarmModelResponse_alarmModelArn - The ARN of the alarm model. For more information, see Amazon Resource Names (ARNs) in the AWS General Reference.

$sel:httpStatus:UpdateAlarmModelResponse', updateAlarmModelResponse_httpStatus - The response's http status code.

Response Lenses

updateAlarmModelResponse_creationTime :: Lens' UpdateAlarmModelResponse (Maybe UTCTime)

The time the alarm model was created, in the Unix epoch format.

updateAlarmModelResponse_status :: Lens' UpdateAlarmModelResponse (Maybe AlarmModelVersionStatus)

The status of the alarm model. The status can be one of the following values:

ACTIVE - The alarm model is active and it's ready to evaluate data.

ACTIVATING - AWS IoT Events is activating your alarm model. Activating an alarm model can take up to a few minutes.

INACTIVE - The alarm model is inactive, so it isn't ready to evaluate data. Check your alarm model information and update the alarm model.

FAILED - You couldn't create or update the alarm model. Check your alarm model information and try again.

updateAlarmModelResponse_lastUpdateTime :: Lens' UpdateAlarmModelResponse (Maybe UTCTime)

The time the alarm model was last updated, in the Unix epoch format.

updateAlarmModelResponse_alarmModelVersion :: Lens' UpdateAlarmModelResponse (Maybe Text)

The version of the alarm model.

updateAlarmModelResponse_alarmModelArn :: Lens' UpdateAlarmModelResponse (Maybe Text)

The ARN of the alarm model. For more information, see Amazon Resource Names (ARNs) in the AWS General Reference.

updateAlarmModelResponse_httpStatus :: Lens' UpdateAlarmModelResponse Int

The response's http status code.