Amazonka.IoTEventsData.BatchAcknowledgeAlarm

Description

Acknowledges one or more alarms. The alarms change to the ACKNOWLEDGED state after you acknowledge them.

Creating a Request

data BatchAcknowledgeAlarm

See: newBatchAcknowledgeAlarm smart constructor.

Constructors

BatchAcknowledgeAlarm'

Fields

acknowledgeActionRequests :: NonEmpty AcknowledgeAlarmActionRequest

The list of acknowledge action requests. You can specify up to 10 requests per operation.

newBatchAcknowledgeAlarm

Arguments

:: NonEmpty AcknowledgeAlarmActionRequest
$sel:acknowledgeActionRequests:BatchAcknowledgeAlarm'

-> BatchAcknowledgeAlarm

Create a value of BatchAcknowledgeAlarm with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:acknowledgeActionRequests:BatchAcknowledgeAlarm', batchAcknowledgeAlarm_acknowledgeActionRequests - The list of acknowledge action requests. You can specify up to 10 requests per operation.

Request Lenses

batchAcknowledgeAlarm_acknowledgeActionRequests :: Lens' BatchAcknowledgeAlarm (NonEmpty AcknowledgeAlarmActionRequest)

The list of acknowledge action requests. You can specify up to 10 requests per operation.

Destructuring the Response

data BatchAcknowledgeAlarmResponse

See: newBatchAcknowledgeAlarmResponse smart constructor.

Constructors

BatchAcknowledgeAlarmResponse'

Fields

errorEntries :: Maybe [BatchAlarmActionErrorEntry]

A list of errors associated with the request, or null if there are no errors. Each error entry contains an entry ID that helps you identify the entry that failed.

httpStatus :: Int

The response's http status code.

newBatchAcknowledgeAlarmResponse

Arguments

:: Int
$sel:httpStatus:BatchAcknowledgeAlarmResponse'

-> BatchAcknowledgeAlarmResponse

Create a value of BatchAcknowledgeAlarmResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:errorEntries:BatchAcknowledgeAlarmResponse', batchAcknowledgeAlarmResponse_errorEntries - A list of errors associated with the request, or null if there are no errors. Each error entry contains an entry ID that helps you identify the entry that failed.

$sel:httpStatus:BatchAcknowledgeAlarmResponse', batchAcknowledgeAlarmResponse_httpStatus - The response's http status code.

Response Lenses

batchAcknowledgeAlarmResponse_errorEntries :: Lens' BatchAcknowledgeAlarmResponse (Maybe [BatchAlarmActionErrorEntry])

A list of errors associated with the request, or null if there are no errors. Each error entry contains an entry ID that helps you identify the entry that failed.

batchAcknowledgeAlarmResponse_httpStatus :: Lens' BatchAcknowledgeAlarmResponse Int

The response's http status code.