Amazonka.IoTEventsData.Lens

Description

Operations

BatchSnoozeAlarm

batchSnoozeAlarm_snoozeActionRequests :: Lens' BatchSnoozeAlarm (NonEmpty SnoozeAlarmActionRequest)

The list of snooze action requests. You can specify up to 10 requests per operation.

batchSnoozeAlarmResponse_errorEntries :: Lens' BatchSnoozeAlarmResponse (Maybe [BatchAlarmActionErrorEntry])

A list of errors associated with the request, or null if there are no errors. Each error entry contains an entry ID that helps you identify the entry that failed.

batchSnoozeAlarmResponse_httpStatus :: Lens' BatchSnoozeAlarmResponse Int

The response's http status code.

BatchDisableAlarm

batchDisableAlarm_disableActionRequests :: Lens' BatchDisableAlarm (NonEmpty DisableAlarmActionRequest)

The list of disable action requests. You can specify up to 10 requests per operation.

batchDisableAlarmResponse_errorEntries :: Lens' BatchDisableAlarmResponse (Maybe [BatchAlarmActionErrorEntry])

A list of errors associated with the request, or null if there are no errors. Each error entry contains an entry ID that helps you identify the entry that failed.

batchDisableAlarmResponse_httpStatus :: Lens' BatchDisableAlarmResponse Int

The response's http status code.

DescribeAlarm

describeAlarm_keyValue :: Lens' DescribeAlarm (Maybe Text)

The value of the key used as a filter to select only the alarms associated with the key.

describeAlarm_alarmModelName :: Lens' DescribeAlarm Text

The name of the alarm model.

describeAlarmResponse_alarm :: Lens' DescribeAlarmResponse (Maybe Alarm)

Contains information about an alarm.

describeAlarmResponse_httpStatus :: Lens' DescribeAlarmResponse Int

The response's http status code.

BatchPutMessage

batchPutMessage_messages :: Lens' BatchPutMessage (NonEmpty Message)

The list of messages to send. Each message has the following format: '{ "messageId": "string", "inputName": "string", "payload": "string"}'

batchPutMessageResponse_batchPutMessageErrorEntries :: Lens' BatchPutMessageResponse (Maybe [BatchPutMessageErrorEntry])

A list of any errors encountered when sending the messages.

batchPutMessageResponse_httpStatus :: Lens' BatchPutMessageResponse Int

The response's http status code.

DescribeDetector

describeDetector_keyValue :: Lens' DescribeDetector (Maybe Text)

A filter used to limit results to detectors (instances) created because of the given key ID.

describeDetector_detectorModelName :: Lens' DescribeDetector Text

The name of the detector model whose detectors (instances) you want information about.

describeDetectorResponse_detector :: Lens' DescribeDetectorResponse (Maybe Detector)

Information about the detector (instance).

describeDetectorResponse_httpStatus :: Lens' DescribeDetectorResponse Int

The response's http status code.

BatchUpdateDetector

batchUpdateDetector_detectors :: Lens' BatchUpdateDetector (NonEmpty UpdateDetectorRequest)

The list of detectors (instances) to update, along with the values to update.

batchUpdateDetectorResponse_batchUpdateDetectorErrorEntries :: Lens' BatchUpdateDetectorResponse (Maybe [BatchUpdateDetectorErrorEntry])

A list of those detector updates that resulted in errors. (If an error is listed here, the specific update did not occur.)

batchUpdateDetectorResponse_httpStatus :: Lens' BatchUpdateDetectorResponse Int

The response's http status code.

BatchAcknowledgeAlarm

batchAcknowledgeAlarm_acknowledgeActionRequests :: Lens' BatchAcknowledgeAlarm (NonEmpty AcknowledgeAlarmActionRequest)

The list of acknowledge action requests. You can specify up to 10 requests per operation.

batchAcknowledgeAlarmResponse_errorEntries :: Lens' BatchAcknowledgeAlarmResponse (Maybe [BatchAlarmActionErrorEntry])

A list of errors associated with the request, or null if there are no errors. Each error entry contains an entry ID that helps you identify the entry that failed.

batchAcknowledgeAlarmResponse_httpStatus :: Lens' BatchAcknowledgeAlarmResponse Int

The response's http status code.

ListAlarms

listAlarms_nextToken :: Lens' ListAlarms (Maybe Text)

The token that you can use to return the next set of results.

listAlarms_maxResults :: Lens' ListAlarms (Maybe Natural)

The maximum number of results to be returned per request.

listAlarms_alarmModelName :: Lens' ListAlarms Text

The name of the alarm model.

listAlarmsResponse_nextToken :: Lens' ListAlarmsResponse (Maybe Text)

The token that you can use to return the next set of results, or null if there are no more results.

listAlarmsResponse_alarmSummaries :: Lens' ListAlarmsResponse (Maybe [AlarmSummary])

A list that summarizes each alarm.

listAlarmsResponse_httpStatus :: Lens' ListAlarmsResponse Int

The response's http status code.

BatchResetAlarm

batchResetAlarm_resetActionRequests :: Lens' BatchResetAlarm (NonEmpty ResetAlarmActionRequest)

The list of reset action requests. You can specify up to 10 requests per operation.

batchResetAlarmResponse_errorEntries :: Lens' BatchResetAlarmResponse (Maybe [BatchAlarmActionErrorEntry])

A list of errors associated with the request, or null if there are no errors. Each error entry contains an entry ID that helps you identify the entry that failed.

batchResetAlarmResponse_httpStatus :: Lens' BatchResetAlarmResponse Int

The response's http status code.

ListDetectors

listDetectors_nextToken :: Lens' ListDetectors (Maybe Text)

The token that you can use to return the next set of results.

listDetectors_stateName :: Lens' ListDetectors (Maybe Text)

A filter that limits results to those detectors (instances) in the given state.

listDetectors_maxResults :: Lens' ListDetectors (Maybe Natural)

The maximum number of results to be returned per request.

listDetectors_detectorModelName :: Lens' ListDetectors Text

The name of the detector model whose detectors (instances) are listed.

listDetectorsResponse_nextToken :: Lens' ListDetectorsResponse (Maybe Text)

The token that you can use to return the next set of results, or null if there are no more results.

listDetectorsResponse_detectorSummaries :: Lens' ListDetectorsResponse (Maybe [DetectorSummary])

A list of summary information about the detectors (instances).

listDetectorsResponse_httpStatus :: Lens' ListDetectorsResponse Int

The response's http status code.

BatchEnableAlarm

batchEnableAlarm_enableActionRequests :: Lens' BatchEnableAlarm (NonEmpty EnableAlarmActionRequest)

The list of enable action requests. You can specify up to 10 requests per operation.

batchEnableAlarmResponse_errorEntries :: Lens' BatchEnableAlarmResponse (Maybe [BatchAlarmActionErrorEntry])

A list of errors associated with the request, or null if there are no errors. Each error entry contains an entry ID that helps you identify the entry that failed.

batchEnableAlarmResponse_httpStatus :: Lens' BatchEnableAlarmResponse Int

The response's http status code.

Types

AcknowledgeActionConfiguration

acknowledgeActionConfiguration_note :: Lens' AcknowledgeActionConfiguration (Maybe Text)

The note that you can leave when you acknowledge the alarm.

AcknowledgeAlarmActionRequest

acknowledgeAlarmActionRequest_keyValue :: Lens' AcknowledgeAlarmActionRequest (Maybe Text)

The value of the key used as a filter to select only the alarms associated with the key.

acknowledgeAlarmActionRequest_note :: Lens' AcknowledgeAlarmActionRequest (Maybe Text)

The note that you can leave when you acknowledge the alarm.

acknowledgeAlarmActionRequest_requestId :: Lens' AcknowledgeAlarmActionRequest Text

The request ID. Each ID must be unique within each batch.

acknowledgeAlarmActionRequest_alarmModelName :: Lens' AcknowledgeAlarmActionRequest Text

The name of the alarm model.

Alarm

alarm_keyValue :: Lens' Alarm (Maybe Text)

The value of the key used as a filter to select only the alarms associated with the key.

alarm_creationTime :: Lens' Alarm (Maybe UTCTime)

The time the alarm was created, in the Unix epoch format.

alarm_alarmState :: Lens' Alarm (Maybe AlarmState)

Contains information about the current state of the alarm.

alarm_alarmModelName :: Lens' Alarm (Maybe Text)

The name of the alarm model.

alarm_severity :: Lens' Alarm (Maybe Natural)

A non-negative integer that reflects the severity level of the alarm.

alarm_lastUpdateTime :: Lens' Alarm (Maybe UTCTime)

The time the alarm was last updated, in the Unix epoch format.

alarm_alarmModelVersion :: Lens' Alarm (Maybe Text)

The version of the alarm model.

AlarmState

alarmState_customerAction :: Lens' AlarmState (Maybe CustomerAction)

Contains information about the action that you can take to respond to the alarm.

alarmState_stateName :: Lens' AlarmState (Maybe AlarmStateName)

The name of the alarm state. The state name can be one of the following values:

DISABLED - When the alarm is in the DISABLED state, it isn't ready to evaluate data. To enable the alarm, you must change the alarm to the NORMAL state.

NORMAL - When the alarm is in the NORMAL state, it's ready to evaluate data.

ACTIVE - If the alarm is in the ACTIVE state, the alarm is invoked.

ACKNOWLEDGED - When the alarm is in the ACKNOWLEDGED state, the alarm was invoked and you acknowledged the alarm.

SNOOZE_DISABLED - When the alarm is in the SNOOZE_DISABLED state, the alarm is disabled for a specified period of time. After the snooze time, the alarm automatically changes to the NORMAL state.

LATCHED - When the alarm is in the LATCHED state, the alarm was invoked. However, the data that the alarm is currently evaluating is within the specified range. To change the alarm to the NORMAL state, you must acknowledge the alarm.

alarmState_ruleEvaluation :: Lens' AlarmState (Maybe RuleEvaluation)

Information needed to evaluate data.

alarmState_systemEvent :: Lens' AlarmState (Maybe SystemEvent)

Contains information about alarm state changes.

AlarmSummary

alarmSummary_keyValue :: Lens' AlarmSummary (Maybe Text)

The value of the key used as a filter to select only the alarms associated with the key.

alarmSummary_creationTime :: Lens' AlarmSummary (Maybe UTCTime)

The time the alarm was created, in the Unix epoch format.

alarmSummary_alarmModelName :: Lens' AlarmSummary (Maybe Text)

The name of the alarm model.

alarmSummary_stateName :: Lens' AlarmSummary (Maybe AlarmStateName)

The name of the alarm state. The state name can be one of the following values:

DISABLED - When the alarm is in the DISABLED state, it isn't ready to evaluate data. To enable the alarm, you must change the alarm to the NORMAL state.

NORMAL - When the alarm is in the NORMAL state, it's ready to evaluate data.

ACTIVE - If the alarm is in the ACTIVE state, the alarm is invoked.

ACKNOWLEDGED - When the alarm is in the ACKNOWLEDGED state, the alarm was invoked and you acknowledged the alarm.

SNOOZE_DISABLED - When the alarm is in the SNOOZE_DISABLED state, the alarm is disabled for a specified period of time. After the snooze time, the alarm automatically changes to the NORMAL state.

LATCHED - When the alarm is in the LATCHED state, the alarm was invoked. However, the data that the alarm is currently evaluating is within the specified range. To change the alarm to the NORMAL state, you must acknowledge the alarm.

alarmSummary_lastUpdateTime :: Lens' AlarmSummary (Maybe UTCTime)

The time the alarm was last updated, in the Unix epoch format.

alarmSummary_alarmModelVersion :: Lens' AlarmSummary (Maybe Text)

The version of the alarm model.

BatchAlarmActionErrorEntry

batchAlarmActionErrorEntry_requestId :: Lens' BatchAlarmActionErrorEntry (Maybe Text)

The request ID. Each ID must be unique within each batch.

batchAlarmActionErrorEntry_errorCode :: Lens' BatchAlarmActionErrorEntry (Maybe ErrorCode)

The error code.

batchAlarmActionErrorEntry_errorMessage :: Lens' BatchAlarmActionErrorEntry (Maybe Text)

A message that describes the error.

BatchPutMessageErrorEntry

batchPutMessageErrorEntry_errorCode :: Lens' BatchPutMessageErrorEntry (Maybe ErrorCode)

The error code.

batchPutMessageErrorEntry_errorMessage :: Lens' BatchPutMessageErrorEntry (Maybe Text)

A message that describes the error.

batchPutMessageErrorEntry_messageId :: Lens' BatchPutMessageErrorEntry (Maybe Text)

The ID of the message that caused the error. (See the value corresponding to the "messageId" key in the "message" object.)

BatchUpdateDetectorErrorEntry

batchUpdateDetectorErrorEntry_errorCode :: Lens' BatchUpdateDetectorErrorEntry (Maybe ErrorCode)

The error code.

batchUpdateDetectorErrorEntry_errorMessage :: Lens' BatchUpdateDetectorErrorEntry (Maybe Text)

A message that describes the error.

batchUpdateDetectorErrorEntry_messageId :: Lens' BatchUpdateDetectorErrorEntry (Maybe Text)

The "messageId" of the update request that caused the error. (The value of the "messageId" in the update request "Detector" object.)

CustomerAction

customerAction_resetActionConfiguration :: Lens' CustomerAction (Maybe ResetActionConfiguration)

Contains the configuration information of a reset action.

customerAction_actionName :: Lens' CustomerAction (Maybe CustomerActionName)

The name of the action. The action name can be one of the following values:

SNOOZE - When you snooze the alarm, the alarm state changes to SNOOZE_DISABLED.

ENABLE - When you enable the alarm, the alarm state changes to NORMAL.

DISABLE - When you disable the alarm, the alarm state changes to DISABLED.

ACKNOWLEDGE - When you acknowledge the alarm, the alarm state changes to ACKNOWLEDGED.

RESET - When you reset the alarm, the alarm state changes to NORMAL.

For more information, see the AlarmState API.

customerAction_snoozeActionConfiguration :: Lens' CustomerAction (Maybe SnoozeActionConfiguration)

Contains the configuration information of a snooze action.

customerAction_enableActionConfiguration :: Lens' CustomerAction (Maybe EnableActionConfiguration)

Contains the configuration information of an enable action.

customerAction_disableActionConfiguration :: Lens' CustomerAction (Maybe DisableActionConfiguration)

Contains the configuration information of a disable action.

customerAction_acknowledgeActionConfiguration :: Lens' CustomerAction (Maybe AcknowledgeActionConfiguration)

Contains the configuration information of an acknowledge action.

Detector

detector_keyValue :: Lens' Detector (Maybe Text)

The value of the key (identifying the device or system) that caused the creation of this detector (instance).

detector_creationTime :: Lens' Detector (Maybe UTCTime)

The time the detector (instance) was created.

detector_state :: Lens' Detector (Maybe DetectorState)

The current state of the detector (instance).

detector_detectorModelName :: Lens' Detector (Maybe Text)

The name of the detector model that created this detector (instance).

detector_detectorModelVersion :: Lens' Detector (Maybe Text)

The version of the detector model that created this detector (instance).

detector_lastUpdateTime :: Lens' Detector (Maybe UTCTime)

The time the detector (instance) was last updated.

DetectorState

detectorState_stateName :: Lens' DetectorState Text

The name of the state.

detectorState_variables :: Lens' DetectorState [Variable]

The current values of the detector's variables.

detectorState_timers :: Lens' DetectorState [Timer]

The current state of the detector's timers.

DetectorStateDefinition

detectorStateDefinition_stateName :: Lens' DetectorStateDefinition Text

The name of the new state of the detector (instance).

detectorStateDefinition_variables :: Lens' DetectorStateDefinition [VariableDefinition]

The new values of the detector's variables. Any variable whose value isn't specified is cleared.

detectorStateDefinition_timers :: Lens' DetectorStateDefinition [TimerDefinition]

The new values of the detector's timers. Any timer whose value isn't specified is cleared, and its timeout event won't occur.

DetectorStateSummary

detectorStateSummary_stateName :: Lens' DetectorStateSummary (Maybe Text)

The name of the state.

DetectorSummary

detectorSummary_keyValue :: Lens' DetectorSummary (Maybe Text)

The value of the key (identifying the device or system) that caused the creation of this detector (instance).

detectorSummary_creationTime :: Lens' DetectorSummary (Maybe UTCTime)

The time the detector (instance) was created.

detectorSummary_state :: Lens' DetectorSummary (Maybe DetectorStateSummary)

The current state of the detector (instance).

detectorSummary_detectorModelName :: Lens' DetectorSummary (Maybe Text)

The name of the detector model that created this detector (instance).

detectorSummary_detectorModelVersion :: Lens' DetectorSummary (Maybe Text)

The version of the detector model that created this detector (instance).

detectorSummary_lastUpdateTime :: Lens' DetectorSummary (Maybe UTCTime)

The time the detector (instance) was last updated.

DisableActionConfiguration

disableActionConfiguration_note :: Lens' DisableActionConfiguration (Maybe Text)

The note that you can leave when you disable the alarm.

DisableAlarmActionRequest

disableAlarmActionRequest_keyValue :: Lens' DisableAlarmActionRequest (Maybe Text)

The value of the key used as a filter to select only the alarms associated with the key.

disableAlarmActionRequest_note :: Lens' DisableAlarmActionRequest (Maybe Text)

The note that you can leave when you disable the alarm.

disableAlarmActionRequest_requestId :: Lens' DisableAlarmActionRequest Text

The request ID. Each ID must be unique within each batch.

disableAlarmActionRequest_alarmModelName :: Lens' DisableAlarmActionRequest Text

The name of the alarm model.

EnableActionConfiguration

enableActionConfiguration_note :: Lens' EnableActionConfiguration (Maybe Text)

The note that you can leave when you enable the alarm.

EnableAlarmActionRequest

enableAlarmActionRequest_keyValue :: Lens' EnableAlarmActionRequest (Maybe Text)

The value of the key used as a filter to select only the alarms associated with the key.

enableAlarmActionRequest_note :: Lens' EnableAlarmActionRequest (Maybe Text)

The note that you can leave when you enable the alarm.

enableAlarmActionRequest_requestId :: Lens' EnableAlarmActionRequest Text

The request ID. Each ID must be unique within each batch.

enableAlarmActionRequest_alarmModelName :: Lens' EnableAlarmActionRequest Text

The name of the alarm model.

Message

message_timestamp :: Lens' Message (Maybe TimestampValue)

The timestamp associated with the message.

message_messageId :: Lens' Message Text

The ID to assign to the message. Within each batch sent, each "messageId" must be unique.

message_inputName :: Lens' Message Text

The name of the input into which the message payload is transformed.

message_payload :: Lens' Message ByteString

The payload of the message. This can be a JSON string or a Base-64-encoded string representing binary data (in which case you must decode it).-- -- Note: This Lens automatically encodes and decodes Base64 data. -- The underlying isomorphism will encode to Base64 representation during -- serialisation, and decode from Base64 representation during deserialisation. -- This Lens accepts and returns only raw unencoded data.

ResetActionConfiguration

resetActionConfiguration_note :: Lens' ResetActionConfiguration (Maybe Text)

The note that you can leave when you reset the alarm.

ResetAlarmActionRequest

resetAlarmActionRequest_keyValue :: Lens' ResetAlarmActionRequest (Maybe Text)

The value of the key used as a filter to select only the alarms associated with the key.

resetAlarmActionRequest_note :: Lens' ResetAlarmActionRequest (Maybe Text)

The note that you can leave when you reset the alarm.

resetAlarmActionRequest_requestId :: Lens' ResetAlarmActionRequest Text

The request ID. Each ID must be unique within each batch.

resetAlarmActionRequest_alarmModelName :: Lens' ResetAlarmActionRequest Text

The name of the alarm model.

RuleEvaluation

ruleEvaluation_simpleRuleEvaluation :: Lens' RuleEvaluation (Maybe SimpleRuleEvaluation)

Information needed to compare two values with a comparison operator.

SimpleRuleEvaluation

simpleRuleEvaluation_thresholdValue :: Lens' SimpleRuleEvaluation (Maybe Text)

The threshold value, on the right side of the comparison operator.

simpleRuleEvaluation_inputPropertyValue :: Lens' SimpleRuleEvaluation (Maybe Text)

The value of the input property, on the left side of the comparison operator.

simpleRuleEvaluation_operator :: Lens' SimpleRuleEvaluation (Maybe ComparisonOperator)

The comparison operator.

SnoozeActionConfiguration

snoozeActionConfiguration_note :: Lens' SnoozeActionConfiguration (Maybe Text)

The note that you can leave when you snooze the alarm.

snoozeActionConfiguration_snoozeDuration :: Lens' SnoozeActionConfiguration (Maybe Int)

The snooze time in seconds. The alarm automatically changes to the NORMAL state after this duration.

SnoozeAlarmActionRequest

snoozeAlarmActionRequest_keyValue :: Lens' SnoozeAlarmActionRequest (Maybe Text)

The value of the key used as a filter to select only the alarms associated with the key.

snoozeAlarmActionRequest_note :: Lens' SnoozeAlarmActionRequest (Maybe Text)

The note that you can leave when you snooze the alarm.

snoozeAlarmActionRequest_requestId :: Lens' SnoozeAlarmActionRequest Text

The request ID. Each ID must be unique within each batch.

snoozeAlarmActionRequest_alarmModelName :: Lens' SnoozeAlarmActionRequest Text

The name of the alarm model.

snoozeAlarmActionRequest_snoozeDuration :: Lens' SnoozeAlarmActionRequest Int

The snooze time in seconds. The alarm automatically changes to the NORMAL state after this duration.

StateChangeConfiguration

stateChangeConfiguration_triggerType :: Lens' StateChangeConfiguration (Maybe TriggerType)

The trigger type. If the value is SNOOZE_TIMEOUT, the snooze duration ends and the alarm automatically changes to the NORMAL state.

SystemEvent

systemEvent_eventType :: Lens' SystemEvent (Maybe EventType)

The event type. If the value is STATE_CHANGE, the event contains information about alarm state changes.

systemEvent_stateChangeConfiguration :: Lens' SystemEvent (Maybe StateChangeConfiguration)

Contains the configuration information of alarm state changes.

Timer

timer_name :: Lens' Timer Text

The name of the timer.

timer_timestamp :: Lens' Timer UTCTime

The number of seconds which have elapsed on the timer.

TimerDefinition

timerDefinition_name :: Lens' TimerDefinition Text

The name of the timer.

timerDefinition_seconds :: Lens' TimerDefinition Int

The new setting of the timer (the number of seconds before the timer elapses).

TimestampValue

timestampValue_timeInMillis :: Lens' TimestampValue (Maybe Natural)

The value of the timestamp, in the Unix epoch format.

UpdateDetectorRequest

updateDetectorRequest_keyValue :: Lens' UpdateDetectorRequest (Maybe Text)

The value of the input key attribute (identifying the device or system) that caused the creation of this detector (instance).

updateDetectorRequest_messageId :: Lens' UpdateDetectorRequest Text

The ID to assign to the detector update "message". Each "messageId" must be unique within each batch sent.

updateDetectorRequest_detectorModelName :: Lens' UpdateDetectorRequest Text

The name of the detector model that created the detectors (instances).

updateDetectorRequest_state :: Lens' UpdateDetectorRequest DetectorStateDefinition

The new state, variable values, and timer settings of the detector (instance).

Variable

variable_name :: Lens' Variable Text

The name of the variable.

variable_value :: Lens' Variable Text

The current value of the variable.

VariableDefinition

variableDefinition_name :: Lens' VariableDefinition Text

The name of the variable.

variableDefinition_value :: Lens' VariableDefinition Text

The new value of the variable.