Amazonka.IoTEventsData.Types.AlarmSummary

Description

Documentation

data AlarmSummary

Contains a summary of an alarm.

See: newAlarmSummary smart constructor.

Constructors

AlarmSummary'

Fields

keyValue :: Maybe Text

The value of the key used as a filter to select only the alarms associated with the key.

creationTime :: Maybe POSIX

The time the alarm was created, in the Unix epoch format.

alarmModelName :: Maybe Text

The name of the alarm model.

stateName :: Maybe AlarmStateName

The name of the alarm state. The state name can be one of the following values:

DISABLED - When the alarm is in the DISABLED state, it isn't ready to evaluate data. To enable the alarm, you must change the alarm to the NORMAL state.

NORMAL - When the alarm is in the NORMAL state, it's ready to evaluate data.

ACTIVE - If the alarm is in the ACTIVE state, the alarm is invoked.

ACKNOWLEDGED - When the alarm is in the ACKNOWLEDGED state, the alarm was invoked and you acknowledged the alarm.

SNOOZE_DISABLED - When the alarm is in the SNOOZE_DISABLED state, the alarm is disabled for a specified period of time. After the snooze time, the alarm automatically changes to the NORMAL state.

LATCHED - When the alarm is in the LATCHED state, the alarm was invoked. However, the data that the alarm is currently evaluating is within the specified range. To change the alarm to the NORMAL state, you must acknowledge the alarm.

lastUpdateTime :: Maybe POSIX

The time the alarm was last updated, in the Unix epoch format.

alarmModelVersion :: Maybe Text

The version of the alarm model.

newAlarmSummary :: AlarmSummary

Create a value of AlarmSummary with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:keyValue:AlarmSummary', alarmSummary_keyValue - The value of the key used as a filter to select only the alarms associated with the key.

$sel:creationTime:AlarmSummary', alarmSummary_creationTime - The time the alarm was created, in the Unix epoch format.

$sel:alarmModelName:AlarmSummary', alarmSummary_alarmModelName - The name of the alarm model.

$sel:stateName:AlarmSummary', alarmSummary_stateName - The name of the alarm state. The state name can be one of the following values:

DISABLED - When the alarm is in the DISABLED state, it isn't ready to evaluate data. To enable the alarm, you must change the alarm to the NORMAL state.

NORMAL - When the alarm is in the NORMAL state, it's ready to evaluate data.

ACTIVE - If the alarm is in the ACTIVE state, the alarm is invoked.

ACKNOWLEDGED - When the alarm is in the ACKNOWLEDGED state, the alarm was invoked and you acknowledged the alarm.

SNOOZE_DISABLED - When the alarm is in the SNOOZE_DISABLED state, the alarm is disabled for a specified period of time. After the snooze time, the alarm automatically changes to the NORMAL state.

LATCHED - When the alarm is in the LATCHED state, the alarm was invoked. However, the data that the alarm is currently evaluating is within the specified range. To change the alarm to the NORMAL state, you must acknowledge the alarm.

$sel:lastUpdateTime:AlarmSummary', alarmSummary_lastUpdateTime - The time the alarm was last updated, in the Unix epoch format.

$sel:alarmModelVersion:AlarmSummary', alarmSummary_alarmModelVersion - The version of the alarm model.

alarmSummary_keyValue :: Lens' AlarmSummary (Maybe Text)

The value of the key used as a filter to select only the alarms associated with the key.

alarmSummary_creationTime :: Lens' AlarmSummary (Maybe UTCTime)

The time the alarm was created, in the Unix epoch format.

alarmSummary_alarmModelName :: Lens' AlarmSummary (Maybe Text)

The name of the alarm model.

alarmSummary_stateName :: Lens' AlarmSummary (Maybe AlarmStateName)

The name of the alarm state. The state name can be one of the following values:

DISABLED - When the alarm is in the DISABLED state, it isn't ready to evaluate data. To enable the alarm, you must change the alarm to the NORMAL state.

NORMAL - When the alarm is in the NORMAL state, it's ready to evaluate data.

ACTIVE - If the alarm is in the ACTIVE state, the alarm is invoked.

ACKNOWLEDGED - When the alarm is in the ACKNOWLEDGED state, the alarm was invoked and you acknowledged the alarm.

SNOOZE_DISABLED - When the alarm is in the SNOOZE_DISABLED state, the alarm is disabled for a specified period of time. After the snooze time, the alarm automatically changes to the NORMAL state.

LATCHED - When the alarm is in the LATCHED state, the alarm was invoked. However, the data that the alarm is currently evaluating is within the specified range. To change the alarm to the NORMAL state, you must acknowledge the alarm.

alarmSummary_lastUpdateTime :: Lens' AlarmSummary (Maybe UTCTime)

The time the alarm was last updated, in the Unix epoch format.

alarmSummary_alarmModelVersion :: Lens' AlarmSummary (Maybe Text)

The version of the alarm model.