Amazonka.IoTEventsData.Types.CustomerAction

Description

Documentation

data CustomerAction

Contains information about the action that you can take to respond to the alarm.

See: newCustomerAction smart constructor.

Constructors

CustomerAction'

Fields

resetActionConfiguration :: Maybe ResetActionConfiguration

Contains the configuration information of a reset action.

actionName :: Maybe CustomerActionName

The name of the action. The action name can be one of the following values:

SNOOZE - When you snooze the alarm, the alarm state changes to SNOOZE_DISABLED.

ENABLE - When you enable the alarm, the alarm state changes to NORMAL.

DISABLE - When you disable the alarm, the alarm state changes to DISABLED.

ACKNOWLEDGE - When you acknowledge the alarm, the alarm state changes to ACKNOWLEDGED.

RESET - When you reset the alarm, the alarm state changes to NORMAL.

For more information, see the AlarmState API.

snoozeActionConfiguration :: Maybe SnoozeActionConfiguration

Contains the configuration information of a snooze action.

enableActionConfiguration :: Maybe EnableActionConfiguration

Contains the configuration information of an enable action.

disableActionConfiguration :: Maybe DisableActionConfiguration

Contains the configuration information of a disable action.

acknowledgeActionConfiguration :: Maybe AcknowledgeActionConfiguration

Contains the configuration information of an acknowledge action.

newCustomerAction :: CustomerAction

Create a value of CustomerAction with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:resetActionConfiguration:CustomerAction', customerAction_resetActionConfiguration - Contains the configuration information of a reset action.

$sel:actionName:CustomerAction', customerAction_actionName - The name of the action. The action name can be one of the following values:

SNOOZE - When you snooze the alarm, the alarm state changes to SNOOZE_DISABLED.

ENABLE - When you enable the alarm, the alarm state changes to NORMAL.

DISABLE - When you disable the alarm, the alarm state changes to DISABLED.

ACKNOWLEDGE - When you acknowledge the alarm, the alarm state changes to ACKNOWLEDGED.

RESET - When you reset the alarm, the alarm state changes to NORMAL.

For more information, see the AlarmState API.

$sel:snoozeActionConfiguration:CustomerAction', customerAction_snoozeActionConfiguration - Contains the configuration information of a snooze action.

$sel:enableActionConfiguration:CustomerAction', customerAction_enableActionConfiguration - Contains the configuration information of an enable action.

$sel:disableActionConfiguration:CustomerAction', customerAction_disableActionConfiguration - Contains the configuration information of a disable action.

$sel:acknowledgeActionConfiguration:CustomerAction', customerAction_acknowledgeActionConfiguration - Contains the configuration information of an acknowledge action.

customerAction_resetActionConfiguration :: Lens' CustomerAction (Maybe ResetActionConfiguration)

Contains the configuration information of a reset action.

customerAction_actionName :: Lens' CustomerAction (Maybe CustomerActionName)

The name of the action. The action name can be one of the following values:

SNOOZE - When you snooze the alarm, the alarm state changes to SNOOZE_DISABLED.

ENABLE - When you enable the alarm, the alarm state changes to NORMAL.

DISABLE - When you disable the alarm, the alarm state changes to DISABLED.

ACKNOWLEDGE - When you acknowledge the alarm, the alarm state changes to ACKNOWLEDGED.

RESET - When you reset the alarm, the alarm state changes to NORMAL.

For more information, see the AlarmState API.

customerAction_snoozeActionConfiguration :: Lens' CustomerAction (Maybe SnoozeActionConfiguration)

Contains the configuration information of a snooze action.

customerAction_enableActionConfiguration :: Lens' CustomerAction (Maybe EnableActionConfiguration)

Contains the configuration information of an enable action.

customerAction_disableActionConfiguration :: Lens' CustomerAction (Maybe DisableActionConfiguration)

Contains the configuration information of a disable action.

customerAction_acknowledgeActionConfiguration :: Lens' CustomerAction (Maybe AcknowledgeActionConfiguration)

Contains the configuration information of an acknowledge action.