Amazonka.IoTEventsData.Types.Message

Description

Documentation

data Message

Information about a message.

See: newMessage smart constructor.

Constructors

Message'

Fields

timestamp :: Maybe TimestampValue

The timestamp associated with the message.

messageId :: Text

The ID to assign to the message. Within each batch sent, each "messageId" must be unique.

inputName :: Text

The name of the input into which the message payload is transformed.

payload :: Base64

The payload of the message. This can be a JSON string or a Base-64-encoded string representing binary data (in which case you must decode it).

newMessage

Arguments

:: Text
$sel:messageId:Message'

-> Text
$sel:inputName:Message'

-> ByteString
$sel:payload:Message'

-> Message

Create a value of Message with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:timestamp:Message', message_timestamp - The timestamp associated with the message.

$sel:messageId:Message', message_messageId - The ID to assign to the message. Within each batch sent, each "messageId" must be unique.

$sel:inputName:Message', message_inputName - The name of the input into which the message payload is transformed.

$sel:payload:Message', message_payload - The payload of the message. This can be a JSON string or a Base-64-encoded string representing binary data (in which case you must decode it).-- -- Note: This Lens automatically encodes and decodes Base64 data. -- The underlying isomorphism will encode to Base64 representation during -- serialisation, and decode from Base64 representation during deserialisation. -- This Lens accepts and returns only raw unencoded data.

message_timestamp :: Lens' Message (Maybe TimestampValue)

The timestamp associated with the message.

message_messageId :: Lens' Message Text

The ID to assign to the message. Within each batch sent, each "messageId" must be unique.

message_inputName :: Lens' Message Text

The name of the input into which the message payload is transformed.

message_payload :: Lens' Message ByteString

The payload of the message. This can be a JSON string or a Base-64-encoded string representing binary data (in which case you must decode it).-- -- Note: This Lens automatically encodes and decodes Base64 data. -- The underlying isomorphism will encode to Base64 representation during -- serialisation, and decode from Base64 representation during deserialisation. -- This Lens accepts and returns only raw unencoded data.