Amazonka.IoTEventsData.Types.SnoozeActionConfiguration

Description

Documentation

data SnoozeActionConfiguration

Contains the configuration information of a snooze action.

See: newSnoozeActionConfiguration smart constructor.

Constructors

SnoozeActionConfiguration'

Fields

note :: Maybe Text

The note that you can leave when you snooze the alarm.

snoozeDuration :: Maybe Int

The snooze time in seconds. The alarm automatically changes to the NORMAL state after this duration.

newSnoozeActionConfiguration :: SnoozeActionConfiguration

Create a value of SnoozeActionConfiguration with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:note:SnoozeActionConfiguration', snoozeActionConfiguration_note - The note that you can leave when you snooze the alarm.

$sel:snoozeDuration:SnoozeActionConfiguration', snoozeActionConfiguration_snoozeDuration - The snooze time in seconds. The alarm automatically changes to the NORMAL state after this duration.

snoozeActionConfiguration_note :: Lens' SnoozeActionConfiguration (Maybe Text)

The note that you can leave when you snooze the alarm.

snoozeActionConfiguration_snoozeDuration :: Lens' SnoozeActionConfiguration (Maybe Int)

The snooze time in seconds. The alarm automatically changes to the NORMAL state after this duration.