Amazonka.IoTEventsData.Types.SnoozeAlarmActionRequest

Description

Documentation

data SnoozeAlarmActionRequest

Information needed to snooze the alarm.

See: newSnoozeAlarmActionRequest smart constructor.

Constructors

SnoozeAlarmActionRequest'

Fields

keyValue :: Maybe Text

The value of the key used as a filter to select only the alarms associated with the key.

note :: Maybe Text

The note that you can leave when you snooze the alarm.

requestId :: Text

The request ID. Each ID must be unique within each batch.

alarmModelName :: Text

The name of the alarm model.

snoozeDuration :: Int

The snooze time in seconds. The alarm automatically changes to the NORMAL state after this duration.

newSnoozeAlarmActionRequest

Arguments

:: Text
$sel:requestId:SnoozeAlarmActionRequest'

-> Text
$sel:alarmModelName:SnoozeAlarmActionRequest'

-> Int
$sel:snoozeDuration:SnoozeAlarmActionRequest'

-> SnoozeAlarmActionRequest

Create a value of SnoozeAlarmActionRequest with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:keyValue:SnoozeAlarmActionRequest', snoozeAlarmActionRequest_keyValue - The value of the key used as a filter to select only the alarms associated with the key.

$sel:note:SnoozeAlarmActionRequest', snoozeAlarmActionRequest_note - The note that you can leave when you snooze the alarm.

$sel:requestId:SnoozeAlarmActionRequest', snoozeAlarmActionRequest_requestId - The request ID. Each ID must be unique within each batch.

$sel:alarmModelName:SnoozeAlarmActionRequest', snoozeAlarmActionRequest_alarmModelName - The name of the alarm model.

$sel:snoozeDuration:SnoozeAlarmActionRequest', snoozeAlarmActionRequest_snoozeDuration - The snooze time in seconds. The alarm automatically changes to the NORMAL state after this duration.

snoozeAlarmActionRequest_keyValue :: Lens' SnoozeAlarmActionRequest (Maybe Text)

The value of the key used as a filter to select only the alarms associated with the key.

snoozeAlarmActionRequest_note :: Lens' SnoozeAlarmActionRequest (Maybe Text)

The note that you can leave when you snooze the alarm.

snoozeAlarmActionRequest_requestId :: Lens' SnoozeAlarmActionRequest Text

The request ID. Each ID must be unique within each batch.

snoozeAlarmActionRequest_alarmModelName :: Lens' SnoozeAlarmActionRequest Text

The name of the alarm model.

snoozeAlarmActionRequest_snoozeDuration :: Lens' SnoozeAlarmActionRequest Int

The snooze time in seconds. The alarm automatically changes to the NORMAL state after this duration.