Amazonka.IoTEventsData.Types.UpdateDetectorRequest

Description

Documentation

data UpdateDetectorRequest

Information used to update the detector (instance).

See: newUpdateDetectorRequest smart constructor.

Constructors

UpdateDetectorRequest'

Fields

keyValue :: Maybe Text

The value of the input key attribute (identifying the device or system) that caused the creation of this detector (instance).

messageId :: Text

The ID to assign to the detector update "message". Each "messageId" must be unique within each batch sent.

detectorModelName :: Text

The name of the detector model that created the detectors (instances).

state :: DetectorStateDefinition

The new state, variable values, and timer settings of the detector (instance).

newUpdateDetectorRequest

Arguments

:: Text
$sel:messageId:UpdateDetectorRequest'

-> Text
$sel:detectorModelName:UpdateDetectorRequest'

-> DetectorStateDefinition
$sel:state:UpdateDetectorRequest'

-> UpdateDetectorRequest

Create a value of UpdateDetectorRequest with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:keyValue:UpdateDetectorRequest', updateDetectorRequest_keyValue - The value of the input key attribute (identifying the device or system) that caused the creation of this detector (instance).

$sel:messageId:UpdateDetectorRequest', updateDetectorRequest_messageId - The ID to assign to the detector update "message". Each "messageId" must be unique within each batch sent.

$sel:detectorModelName:UpdateDetectorRequest', updateDetectorRequest_detectorModelName - The name of the detector model that created the detectors (instances).

$sel:state:UpdateDetectorRequest', updateDetectorRequest_state - The new state, variable values, and timer settings of the detector (instance).

updateDetectorRequest_keyValue :: Lens' UpdateDetectorRequest (Maybe Text)

The value of the input key attribute (identifying the device or system) that caused the creation of this detector (instance).

updateDetectorRequest_messageId :: Lens' UpdateDetectorRequest Text

The ID to assign to the detector update "message". Each "messageId" must be unique within each batch sent.

updateDetectorRequest_detectorModelName :: Lens' UpdateDetectorRequest Text

The name of the detector model that created the detectors (instances).

updateDetectorRequest_state :: Lens' UpdateDetectorRequest DetectorStateDefinition

The new state, variable values, and timer settings of the detector (instance).