Amazonka.IoTFleetHub.DescribeApplication

Description

Gets information about a Fleet Hub for AWS IoT Device Management web application.

Fleet Hub for AWS IoT Device Management is in public preview and is subject to change.

Creating a Request

data DescribeApplication

See: newDescribeApplication smart constructor.

Constructors

DescribeApplication'

Fields

applicationId :: Text

The unique Id of the web application.

newDescribeApplication

Arguments

:: Text
$sel:applicationId:DescribeApplication'

-> DescribeApplication

Create a value of DescribeApplication with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:applicationId:DescribeApplication', describeApplication_applicationId - The unique Id of the web application.

Request Lenses

describeApplication_applicationId :: Lens' DescribeApplication Text

The unique Id of the web application.

Destructuring the Response

data DescribeApplicationResponse

See: newDescribeApplicationResponse smart constructor.

Constructors

DescribeApplicationResponse'

Fields

applicationDescription :: Maybe Text

An optional description of the web application.

ssoClientId :: Maybe Text

The Id of the single sign-on client that you use to authenticate and authorize users on the web application.

errorMessage :: Maybe Text

A message indicating why the DescribeApplication API failed.

tags :: Maybe (HashMap Text Text)

A set of key/value pairs that you can use to manage the web application resource.

httpStatus :: Int

The response's http status code.

applicationId :: Text

The unique Id of the web application.

applicationArn :: Text

The ARN of the web application.

applicationName :: Text

The name of the web application.

applicationUrl :: Text

The URL of the web application.

applicationState :: ApplicationState

The current state of the web application.

applicationCreationDate :: Integer

The date (in Unix epoch time) when the application was created.

applicationLastUpdateDate :: Integer

The date (in Unix epoch time) when the application was last updated.

roleArn :: Text

The ARN of the role that the web application assumes when it interacts with AWS IoT Core.

newDescribeApplicationResponse

Arguments

:: Int
$sel:httpStatus:DescribeApplicationResponse'

-> Text
$sel:applicationId:DescribeApplication'

-> Text
$sel:applicationArn:DescribeApplicationResponse'

-> Text
$sel:applicationName:DescribeApplicationResponse'

-> Text
$sel:applicationUrl:DescribeApplicationResponse'

-> ApplicationState
$sel:applicationState:DescribeApplicationResponse'

-> Integer
$sel:applicationCreationDate:DescribeApplicationResponse'

-> Integer
$sel:applicationLastUpdateDate:DescribeApplicationResponse'

-> Text
$sel:roleArn:DescribeApplicationResponse'

-> DescribeApplicationResponse

Create a value of DescribeApplicationResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:applicationDescription:DescribeApplicationResponse', describeApplicationResponse_applicationDescription - An optional description of the web application.

$sel:ssoClientId:DescribeApplicationResponse', describeApplicationResponse_ssoClientId - The Id of the single sign-on client that you use to authenticate and authorize users on the web application.

$sel:errorMessage:DescribeApplicationResponse', describeApplicationResponse_errorMessage - A message indicating why the DescribeApplication API failed.

$sel:tags:DescribeApplicationResponse', describeApplicationResponse_tags - A set of key/value pairs that you can use to manage the web application resource.

$sel:httpStatus:DescribeApplicationResponse', describeApplicationResponse_httpStatus - The response's http status code.

$sel:applicationId:DescribeApplication', describeApplicationResponse_applicationId - The unique Id of the web application.

$sel:applicationArn:DescribeApplicationResponse', describeApplicationResponse_applicationArn - The ARN of the web application.

$sel:applicationName:DescribeApplicationResponse', describeApplicationResponse_applicationName - The name of the web application.

$sel:applicationUrl:DescribeApplicationResponse', describeApplicationResponse_applicationUrl - The URL of the web application.

$sel:applicationState:DescribeApplicationResponse', describeApplicationResponse_applicationState - The current state of the web application.

$sel:applicationCreationDate:DescribeApplicationResponse', describeApplicationResponse_applicationCreationDate - The date (in Unix epoch time) when the application was created.

$sel:applicationLastUpdateDate:DescribeApplicationResponse', describeApplicationResponse_applicationLastUpdateDate - The date (in Unix epoch time) when the application was last updated.

$sel:roleArn:DescribeApplicationResponse', describeApplicationResponse_roleArn - The ARN of the role that the web application assumes when it interacts with AWS IoT Core.

Response Lenses

describeApplicationResponse_applicationDescription :: Lens' DescribeApplicationResponse (Maybe Text)

An optional description of the web application.

describeApplicationResponse_ssoClientId :: Lens' DescribeApplicationResponse (Maybe Text)

The Id of the single sign-on client that you use to authenticate and authorize users on the web application.

describeApplicationResponse_errorMessage :: Lens' DescribeApplicationResponse (Maybe Text)

A message indicating why the DescribeApplication API failed.

describeApplicationResponse_tags :: Lens' DescribeApplicationResponse (Maybe (HashMap Text Text))

A set of key/value pairs that you can use to manage the web application resource.

describeApplicationResponse_httpStatus :: Lens' DescribeApplicationResponse Int

The response's http status code.

describeApplicationResponse_applicationId :: Lens' DescribeApplicationResponse Text

The unique Id of the web application.

describeApplicationResponse_applicationArn :: Lens' DescribeApplicationResponse Text

The ARN of the web application.

describeApplicationResponse_applicationName :: Lens' DescribeApplicationResponse Text

The name of the web application.

describeApplicationResponse_applicationUrl :: Lens' DescribeApplicationResponse Text

The URL of the web application.

describeApplicationResponse_applicationState :: Lens' DescribeApplicationResponse ApplicationState

The current state of the web application.

describeApplicationResponse_applicationCreationDate :: Lens' DescribeApplicationResponse Integer

The date (in Unix epoch time) when the application was created.

describeApplicationResponse_applicationLastUpdateDate :: Lens' DescribeApplicationResponse Integer

The date (in Unix epoch time) when the application was last updated.

describeApplicationResponse_roleArn :: Lens' DescribeApplicationResponse Text

The ARN of the role that the web application assumes when it interacts with AWS IoT Core.