Amazonka.IoTFleetHub.Lens

Description

Operations

DescribeApplication

describeApplication_applicationId :: Lens' DescribeApplication Text

The unique Id of the web application.

describeApplicationResponse_applicationDescription :: Lens' DescribeApplicationResponse (Maybe Text)

An optional description of the web application.

describeApplicationResponse_ssoClientId :: Lens' DescribeApplicationResponse (Maybe Text)

The Id of the single sign-on client that you use to authenticate and authorize users on the web application.

describeApplicationResponse_errorMessage :: Lens' DescribeApplicationResponse (Maybe Text)

A message indicating why the DescribeApplication API failed.

describeApplicationResponse_tags :: Lens' DescribeApplicationResponse (Maybe (HashMap Text Text))

A set of key/value pairs that you can use to manage the web application resource.

describeApplicationResponse_httpStatus :: Lens' DescribeApplicationResponse Int

The response's http status code.

describeApplicationResponse_applicationId :: Lens' DescribeApplicationResponse Text

The unique Id of the web application.

describeApplicationResponse_applicationArn :: Lens' DescribeApplicationResponse Text

The ARN of the web application.

describeApplicationResponse_applicationName :: Lens' DescribeApplicationResponse Text

The name of the web application.

describeApplicationResponse_applicationUrl :: Lens' DescribeApplicationResponse Text

The URL of the web application.

describeApplicationResponse_applicationState :: Lens' DescribeApplicationResponse ApplicationState

The current state of the web application.

describeApplicationResponse_applicationCreationDate :: Lens' DescribeApplicationResponse Integer

The date (in Unix epoch time) when the application was created.

describeApplicationResponse_applicationLastUpdateDate :: Lens' DescribeApplicationResponse Integer

The date (in Unix epoch time) when the application was last updated.

describeApplicationResponse_roleArn :: Lens' DescribeApplicationResponse Text

The ARN of the role that the web application assumes when it interacts with AWS IoT Core.

ListTagsForResource

listTagsForResource_resourceArn :: Lens' ListTagsForResource Text

The ARN of the resource.

listTagsForResourceResponse_tags :: Lens' ListTagsForResourceResponse (Maybe (HashMap Text Text))

The list of tags assigned to the resource.

listTagsForResourceResponse_httpStatus :: Lens' ListTagsForResourceResponse Int

The response's http status code.

DeleteApplication

deleteApplication_clientToken :: Lens' DeleteApplication (Maybe Text)

A unique case-sensitive identifier that you can provide to ensure the idempotency of the request. Don't reuse this client token if a new idempotent request is required.

deleteApplication_applicationId :: Lens' DeleteApplication Text

The unique Id of the web application.

deleteApplicationResponse_httpStatus :: Lens' DeleteApplicationResponse Int

The response's http status code.

UpdateApplication

updateApplication_applicationDescription :: Lens' UpdateApplication (Maybe Text)

An optional description of the web application.

updateApplication_clientToken :: Lens' UpdateApplication (Maybe Text)

A unique case-sensitive identifier that you can provide to ensure the idempotency of the request. Don't reuse this client token if a new idempotent request is required.

updateApplication_applicationName :: Lens' UpdateApplication (Maybe Text)

The name of the web application.

updateApplication_applicationId :: Lens' UpdateApplication Text

The unique Id of the web application.

updateApplicationResponse_httpStatus :: Lens' UpdateApplicationResponse Int

The response's http status code.

CreateApplication

createApplication_applicationDescription :: Lens' CreateApplication (Maybe Text)

An optional description of the web application.

createApplication_clientToken :: Lens' CreateApplication (Maybe Text)

A unique case-sensitive identifier that you can provide to ensure the idempotency of the request. Don't reuse this client token if a new idempotent request is required.

createApplication_tags :: Lens' CreateApplication (Maybe (HashMap Text Text))

A set of key/value pairs that you can use to manage the web application resource.

createApplication_applicationName :: Lens' CreateApplication Text

The name of the web application.

createApplication_roleArn :: Lens' CreateApplication Text

The ARN of the role that the web application assumes when it interacts with AWS IoT Core.

The name of the role must be in the form AWSIotFleetHub_random_string .

createApplicationResponse_httpStatus :: Lens' CreateApplicationResponse Int

The response's http status code.

createApplicationResponse_applicationId :: Lens' CreateApplicationResponse Text

The unique Id of the web application.

createApplicationResponse_applicationArn :: Lens' CreateApplicationResponse Text

The ARN of the web application.

TagResource

tagResource_resourceArn :: Lens' TagResource Text

The ARN of the resource.

tagResource_tags :: Lens' TagResource (HashMap Text Text)

The new or modified tags for the resource.

tagResourceResponse_httpStatus :: Lens' TagResourceResponse Int

The response's http status code.

ListApplications

listApplications_nextToken :: Lens' ListApplications (Maybe Text)

A token used to get the next set of results.

listApplicationsResponse_nextToken :: Lens' ListApplicationsResponse (Maybe Text)

A token used to get the next set of results.

listApplicationsResponse_applicationSummaries :: Lens' ListApplicationsResponse (Maybe [ApplicationSummary])

An array of objects that provide summaries of information about the web applications in the list.

listApplicationsResponse_httpStatus :: Lens' ListApplicationsResponse Int

The response's http status code.

UntagResource

untagResource_resourceArn :: Lens' UntagResource Text

The ARN of the resource.

untagResource_tagKeys :: Lens' UntagResource [Text]

A list of the keys of the tags to be removed from the resource.

untagResourceResponse_httpStatus :: Lens' UntagResourceResponse Int

The response's http status code.

Types

ApplicationSummary

applicationSummary_applicationDescription :: Lens' ApplicationSummary (Maybe Text)

An optional description of the web application.

applicationSummary_applicationState :: Lens' ApplicationSummary (Maybe ApplicationState)

The current state of the web application.

applicationSummary_applicationCreationDate :: Lens' ApplicationSummary (Maybe Integer)

The date (in Unix epoch time) when the web application was created.

applicationSummary_applicationLastUpdateDate :: Lens' ApplicationSummary (Maybe Integer)

The date (in Unix epoch time) when the web application was last updated.

applicationSummary_applicationId :: Lens' ApplicationSummary Text

The unique Id of the web application.

applicationSummary_applicationName :: Lens' ApplicationSummary Text

The name of the web application.

applicationSummary_applicationUrl :: Lens' ApplicationSummary Text

The URL of the web application.