Amazonka.IoTJobsData.GetPendingJobExecutions

Description

Gets the list of all jobs for a thing that are not in a terminal status.

Creating a Request

data GetPendingJobExecutions

See: newGetPendingJobExecutions smart constructor.

Constructors

GetPendingJobExecutions'

Fields

thingName :: Text

The name of the thing that is executing the job.

newGetPendingJobExecutions

Arguments

:: Text
$sel:thingName:GetPendingJobExecutions'

-> GetPendingJobExecutions

Create a value of GetPendingJobExecutions with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:thingName:GetPendingJobExecutions', getPendingJobExecutions_thingName - The name of the thing that is executing the job.

Request Lenses

getPendingJobExecutions_thingName :: Lens' GetPendingJobExecutions Text

The name of the thing that is executing the job.

Destructuring the Response

data GetPendingJobExecutionsResponse

See: newGetPendingJobExecutionsResponse smart constructor.

Constructors

GetPendingJobExecutionsResponse'

Fields

inProgressJobs :: Maybe [JobExecutionSummary]

A list of JobExecutionSummary objects with status IN_PROGRESS.

queuedJobs :: Maybe [JobExecutionSummary]

A list of JobExecutionSummary objects with status QUEUED.

httpStatus :: Int

The response's http status code.

newGetPendingJobExecutionsResponse

Arguments

:: Int
$sel:httpStatus:GetPendingJobExecutionsResponse'

-> GetPendingJobExecutionsResponse

Create a value of GetPendingJobExecutionsResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:inProgressJobs:GetPendingJobExecutionsResponse', getPendingJobExecutionsResponse_inProgressJobs - A list of JobExecutionSummary objects with status IN_PROGRESS.

$sel:queuedJobs:GetPendingJobExecutionsResponse', getPendingJobExecutionsResponse_queuedJobs - A list of JobExecutionSummary objects with status QUEUED.

$sel:httpStatus:GetPendingJobExecutionsResponse', getPendingJobExecutionsResponse_httpStatus - The response's http status code.

Response Lenses

getPendingJobExecutionsResponse_inProgressJobs :: Lens' GetPendingJobExecutionsResponse (Maybe [JobExecutionSummary])

A list of JobExecutionSummary objects with status IN_PROGRESS.

getPendingJobExecutionsResponse_queuedJobs :: Lens' GetPendingJobExecutionsResponse (Maybe [JobExecutionSummary])

A list of JobExecutionSummary objects with status QUEUED.

getPendingJobExecutionsResponse_httpStatus :: Lens' GetPendingJobExecutionsResponse Int

The response's http status code.