Amazonka.IoTJobsData.Lens

Description

Operations

UpdateJobExecution

updateJobExecution_includeJobDocument :: Lens' UpdateJobExecution (Maybe Bool)

Optional. When set to true, the response contains the job document. The default is false.

updateJobExecution_stepTimeoutInMinutes :: Lens' UpdateJobExecution (Maybe Integer)

Specifies the amount of time this device has to finish execution of this job. If the job execution status is not set to a terminal state before this timer expires, or before the timer is reset (by again calling UpdateJobExecution, setting the status to IN_PROGRESS and specifying a new timeout value in this field) the job execution status will be automatically set to TIMED_OUT. Note that setting or resetting this timeout has no effect on that job execution timeout which may have been specified when the job was created (CreateJob using field timeoutConfig).

updateJobExecution_statusDetails :: Lens' UpdateJobExecution (Maybe (HashMap Text Text))

Optional. A collection of name/value pairs that describe the status of the job execution. If not specified, the statusDetails are unchanged.

updateJobExecution_executionNumber :: Lens' UpdateJobExecution (Maybe Integer)

Optional. A number that identifies a particular job execution on a particular device.

updateJobExecution_expectedVersion :: Lens' UpdateJobExecution (Maybe Integer)

Optional. The expected current version of the job execution. Each time you update the job execution, its version is incremented. If the version of the job execution stored in Jobs does not match, the update is rejected with a VersionMismatch error, and an ErrorResponse that contains the current job execution status data is returned. (This makes it unnecessary to perform a separate DescribeJobExecution request in order to obtain the job execution status data.)

updateJobExecution_includeJobExecutionState :: Lens' UpdateJobExecution (Maybe Bool)

Optional. When included and set to true, the response contains the JobExecutionState data. The default is false.

updateJobExecution_jobId :: Lens' UpdateJobExecution Text

The unique identifier assigned to this job when it was created.

updateJobExecution_thingName :: Lens' UpdateJobExecution Text

The name of the thing associated with the device.

updateJobExecution_status :: Lens' UpdateJobExecution JobExecutionStatus

The new status for the job execution (IN_PROGRESS, FAILED, SUCCESS, or REJECTED). This must be specified on every update.

updateJobExecutionResponse_jobDocument :: Lens' UpdateJobExecutionResponse (Maybe Text)

The contents of the Job Documents.

updateJobExecutionResponse_executionState :: Lens' UpdateJobExecutionResponse (Maybe JobExecutionState)

A JobExecutionState object.

updateJobExecutionResponse_httpStatus :: Lens' UpdateJobExecutionResponse Int

The response's http status code.

StartNextPendingJobExecution

startNextPendingJobExecution_stepTimeoutInMinutes :: Lens' StartNextPendingJobExecution (Maybe Integer)

Specifies the amount of time this device has to finish execution of this job. If the job execution status is not set to a terminal state before this timer expires, or before the timer is reset (by calling UpdateJobExecution, setting the status to IN_PROGRESS and specifying a new timeout value in field stepTimeoutInMinutes) the job execution status will be automatically set to TIMED_OUT. Note that setting this timeout has no effect on that job execution timeout which may have been specified when the job was created (CreateJob using field timeoutConfig).

startNextPendingJobExecution_statusDetails :: Lens' StartNextPendingJobExecution (Maybe (HashMap Text Text))

A collection of name/value pairs that describe the status of the job execution. If not specified, the statusDetails are unchanged.

startNextPendingJobExecution_thingName :: Lens' StartNextPendingJobExecution Text

The name of the thing associated with the device.

startNextPendingJobExecutionResponse_execution :: Lens' StartNextPendingJobExecutionResponse (Maybe JobExecution)

A JobExecution object.

startNextPendingJobExecutionResponse_httpStatus :: Lens' StartNextPendingJobExecutionResponse Int

The response's http status code.

DescribeJobExecution

describeJobExecution_includeJobDocument :: Lens' DescribeJobExecution (Maybe Bool)

Optional. When set to true, the response contains the job document. The default is false.

describeJobExecution_executionNumber :: Lens' DescribeJobExecution (Maybe Integer)

Optional. A number that identifies a particular job execution on a particular device. If not specified, the latest job execution is returned.

describeJobExecution_jobId :: Lens' DescribeJobExecution Text

The unique identifier assigned to this job when it was created.

describeJobExecution_thingName :: Lens' DescribeJobExecution Text

The thing name associated with the device the job execution is running on.

describeJobExecutionResponse_execution :: Lens' DescribeJobExecutionResponse (Maybe JobExecution)

Contains data about a job execution.

describeJobExecutionResponse_httpStatus :: Lens' DescribeJobExecutionResponse Int

The response's http status code.

GetPendingJobExecutions

getPendingJobExecutions_thingName :: Lens' GetPendingJobExecutions Text

The name of the thing that is executing the job.

getPendingJobExecutionsResponse_inProgressJobs :: Lens' GetPendingJobExecutionsResponse (Maybe [JobExecutionSummary])

A list of JobExecutionSummary objects with status IN_PROGRESS.

getPendingJobExecutionsResponse_queuedJobs :: Lens' GetPendingJobExecutionsResponse (Maybe [JobExecutionSummary])

A list of JobExecutionSummary objects with status QUEUED.

getPendingJobExecutionsResponse_httpStatus :: Lens' GetPendingJobExecutionsResponse Int

The response's http status code.

Types

JobExecution

jobExecution_status :: Lens' JobExecution (Maybe JobExecutionStatus)

The status of the job execution. Can be one of: "QUEUED", "IN_PROGRESS", "FAILED", "SUCCESS", "CANCELED", "REJECTED", or "REMOVED".

jobExecution_jobId :: Lens' JobExecution (Maybe Text)

The unique identifier you assigned to this job when it was created.

jobExecution_lastUpdatedAt :: Lens' JobExecution (Maybe Integer)

The time, in milliseconds since the epoch, when the job execution was last updated.

jobExecution_approximateSecondsBeforeTimedOut :: Lens' JobExecution (Maybe Integer)

The estimated number of seconds that remain before the job execution status will be changed to TIMED_OUT.

jobExecution_queuedAt :: Lens' JobExecution (Maybe Integer)

The time, in milliseconds since the epoch, when the job execution was enqueued.

jobExecution_jobDocument :: Lens' JobExecution (Maybe Text)

The content of the job document.

jobExecution_statusDetails :: Lens' JobExecution (Maybe (HashMap Text Text))

A collection of name/value pairs that describe the status of the job execution.

jobExecution_executionNumber :: Lens' JobExecution (Maybe Integer)

A number that identifies a particular job execution on a particular device. It can be used later in commands that return or update job execution information.

jobExecution_versionNumber :: Lens' JobExecution (Maybe Integer)

The version of the job execution. Job execution versions are incremented each time they are updated by a device.

jobExecution_startedAt :: Lens' JobExecution (Maybe Integer)

The time, in milliseconds since the epoch, when the job execution was started.

jobExecution_thingName :: Lens' JobExecution (Maybe Text)

The name of the thing that is executing the job.

JobExecutionState

jobExecutionState_status :: Lens' JobExecutionState (Maybe JobExecutionStatus)

The status of the job execution. Can be one of: "QUEUED", "IN_PROGRESS", "FAILED", "SUCCESS", "CANCELED", "REJECTED", or "REMOVED".

jobExecutionState_statusDetails :: Lens' JobExecutionState (Maybe (HashMap Text Text))

A collection of name/value pairs that describe the status of the job execution.

jobExecutionState_versionNumber :: Lens' JobExecutionState (Maybe Integer)

The version of the job execution. Job execution versions are incremented each time they are updated by a device.

JobExecutionSummary

jobExecutionSummary_jobId :: Lens' JobExecutionSummary (Maybe Text)

The unique identifier you assigned to this job when it was created.

jobExecutionSummary_lastUpdatedAt :: Lens' JobExecutionSummary (Maybe Integer)

The time, in milliseconds since the epoch, when the job execution was last updated.

jobExecutionSummary_queuedAt :: Lens' JobExecutionSummary (Maybe Integer)

The time, in milliseconds since the epoch, when the job execution was enqueued.

jobExecutionSummary_executionNumber :: Lens' JobExecutionSummary (Maybe Integer)

A number that identifies a particular job execution on a particular device.

jobExecutionSummary_versionNumber :: Lens' JobExecutionSummary (Maybe Integer)

The version of the job execution. Job execution versions are incremented each time AWS IoT Jobs receives an update from a device.

jobExecutionSummary_startedAt :: Lens' JobExecutionSummary (Maybe Integer)

The time, in milliseconds since the epoch, when the job execution started.