Amazonka.IoTSecureTunneling.Lens

Description

Operations

ListTagsForResource

listTagsForResource_resourceArn :: Lens' ListTagsForResource Text

The resource ARN.

listTagsForResourceResponse_tags :: Lens' ListTagsForResourceResponse (Maybe (NonEmpty Tag))

The tags for the specified resource.

listTagsForResourceResponse_httpStatus :: Lens' ListTagsForResourceResponse Int

The response's http status code.

CloseTunnel

closeTunnel_delete :: Lens' CloseTunnel (Maybe Bool)

When set to true, AWS IoT Secure Tunneling deletes the tunnel data immediately.

closeTunnel_tunnelId :: Lens' CloseTunnel Text

The ID of the tunnel to close.

closeTunnelResponse_httpStatus :: Lens' CloseTunnelResponse Int

The response's http status code.

OpenTunnel

openTunnel_destinationConfig :: Lens' OpenTunnel (Maybe DestinationConfig)

The destination configuration for the OpenTunnel request.

openTunnel_description :: Lens' OpenTunnel (Maybe Text)

A short text description of the tunnel.

openTunnel_timeoutConfig :: Lens' OpenTunnel (Maybe TimeoutConfig)

Timeout configuration for a tunnel.

openTunnel_tags :: Lens' OpenTunnel (Maybe (NonEmpty Tag))

A collection of tag metadata.

openTunnelResponse_sourceAccessToken :: Lens' OpenTunnelResponse (Maybe Text)

The access token the source local proxy uses to connect to AWS IoT Secure Tunneling.

openTunnelResponse_tunnelArn :: Lens' OpenTunnelResponse (Maybe Text)

The Amazon Resource Name for the tunnel. The tunnel ARN format is arn:aws:tunnel:<region>:<account-id>:tunnel/<tunnel-id>

openTunnelResponse_destinationAccessToken :: Lens' OpenTunnelResponse (Maybe Text)

The access token the destination local proxy uses to connect to AWS IoT Secure Tunneling.

openTunnelResponse_tunnelId :: Lens' OpenTunnelResponse (Maybe Text)

A unique alpha-numeric tunnel ID.

openTunnelResponse_httpStatus :: Lens' OpenTunnelResponse Int

The response's http status code.

DescribeTunnel

describeTunnel_tunnelId :: Lens' DescribeTunnel Text

The tunnel to describe.

describeTunnelResponse_tunnel :: Lens' DescribeTunnelResponse (Maybe Tunnel)

The tunnel being described.

describeTunnelResponse_httpStatus :: Lens' DescribeTunnelResponse Int

The response's http status code.

TagResource

tagResource_resourceArn :: Lens' TagResource Text

The ARN of the resource.

tagResource_tags :: Lens' TagResource (NonEmpty Tag)

The tags for the resource.

tagResourceResponse_httpStatus :: Lens' TagResourceResponse Int

The response's http status code.

UntagResource

untagResource_resourceArn :: Lens' UntagResource Text

The resource ARN.

untagResource_tagKeys :: Lens' UntagResource [Text]

The keys of the tags to remove.

untagResourceResponse_httpStatus :: Lens' UntagResourceResponse Int

The response's http status code.

ListTunnels

listTunnels_nextToken :: Lens' ListTunnels (Maybe Text)

A token to retrieve the next set of results.

listTunnels_thingName :: Lens' ListTunnels (Maybe Text)

The name of the IoT thing associated with the destination device.

listTunnels_maxResults :: Lens' ListTunnels (Maybe Natural)

The maximum number of results to return at once.

listTunnelsResponse_nextToken :: Lens' ListTunnelsResponse (Maybe Text)

A token to used to retrieve the next set of results.

listTunnelsResponse_tunnelSummaries :: Lens' ListTunnelsResponse (Maybe [TunnelSummary])

A short description of the tunnels in an AWS account.

listTunnelsResponse_httpStatus :: Lens' ListTunnelsResponse Int

The response's http status code.

Types

ConnectionState

connectionState_status :: Lens' ConnectionState (Maybe ConnectionStatus)

The connection status of the tunnel. Valid values are CONNECTED and DISCONNECTED.

connectionState_lastUpdatedAt :: Lens' ConnectionState (Maybe UTCTime)

The last time the connection status was updated.

DestinationConfig

destinationConfig_thingName :: Lens' DestinationConfig (Maybe Text)

The name of the IoT thing to which you want to connect.

destinationConfig_services :: Lens' DestinationConfig (NonEmpty Text)

A list of service names that identity the target application. The AWS IoT client running on the destination device reads this value and uses it to look up a port or an IP address and a port. The AWS IoT client instantiates the local proxy which uses this information to connect to the destination application.

Tag

tag_key :: Lens' Tag Text

The key of the tag.

tag_value :: Lens' Tag Text

The value of the tag.

TimeoutConfig

timeoutConfig_maxLifetimeTimeoutMinutes :: Lens' TimeoutConfig (Maybe Natural)

The maximum amount of time (in minutes) a tunnel can remain open. If not specified, maxLifetimeTimeoutMinutes defaults to 720 minutes. Valid values are from 1 minute to 12 hours (720 minutes)

Tunnel

tunnel_status :: Lens' Tunnel (Maybe TunnelStatus)

The status of a tunnel. Valid values are: Open and Closed.

tunnel_lastUpdatedAt :: Lens' Tunnel (Maybe UTCTime)

The last time the tunnel was updated.

tunnel_createdAt :: Lens' Tunnel (Maybe UTCTime)

The time when the tunnel was created.

tunnel_tunnelArn :: Lens' Tunnel (Maybe Text)

The Amazon Resource Name (ARN) of a tunnel. The tunnel ARN format is arn:aws:tunnel:<region>:<account-id>:tunnel/<tunnel-id>

tunnel_sourceConnectionState :: Lens' Tunnel (Maybe ConnectionState)

The connection state of the source application.

tunnel_destinationConnectionState :: Lens' Tunnel (Maybe ConnectionState)

The connection state of the destination application.

tunnel_tunnelId :: Lens' Tunnel (Maybe Text)

A unique alpha-numeric ID that identifies a tunnel.

tunnel_destinationConfig :: Lens' Tunnel (Maybe DestinationConfig)

The destination configuration that specifies the thing name of the destination device and a service name that the local proxy uses to connect to the destination application.

tunnel_description :: Lens' Tunnel (Maybe Text)

A description of the tunnel.

tunnel_timeoutConfig :: Lens' Tunnel (Maybe TimeoutConfig)

Timeout configuration for the tunnel.

tunnel_tags :: Lens' Tunnel (Maybe (NonEmpty Tag))

A list of tag metadata associated with the secure tunnel.

TunnelSummary

tunnelSummary_status :: Lens' TunnelSummary (Maybe TunnelStatus)

The status of a tunnel. Valid values are: Open and Closed.

tunnelSummary_lastUpdatedAt :: Lens' TunnelSummary (Maybe UTCTime)

The time the tunnel was last updated.

tunnelSummary_createdAt :: Lens' TunnelSummary (Maybe UTCTime)

The time the tunnel was created.

tunnelSummary_tunnelArn :: Lens' TunnelSummary (Maybe Text)

The Amazon Resource Name of the tunnel. The tunnel ARN format is arn:aws:tunnel:<region>:<account-id>:tunnel/<tunnel-id>

tunnelSummary_tunnelId :: Lens' TunnelSummary (Maybe Text)

The unique alpha-numeric identifier for the tunnel.

tunnelSummary_description :: Lens' TunnelSummary (Maybe Text)

A description of the tunnel.