Amazonka.IoTSecureTunneling.ListTunnels

Description

List all tunnels for an AWS account. Tunnels are listed by creation time in descending order, newer tunnels will be listed before older tunnels.

Creating a Request

data ListTunnels

See: newListTunnels smart constructor.

Constructors

ListTunnels'

Fields

nextToken :: Maybe Text

A token to retrieve the next set of results.

thingName :: Maybe Text

The name of the IoT thing associated with the destination device.

maxResults :: Maybe Natural

The maximum number of results to return at once.

newListTunnels :: ListTunnels

Create a value of ListTunnels with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:nextToken:ListTunnels', listTunnels_nextToken - A token to retrieve the next set of results.

$sel:thingName:ListTunnels', listTunnels_thingName - The name of the IoT thing associated with the destination device.

$sel:maxResults:ListTunnels', listTunnels_maxResults - The maximum number of results to return at once.

Request Lenses

listTunnels_nextToken :: Lens' ListTunnels (Maybe Text)

A token to retrieve the next set of results.

listTunnels_thingName :: Lens' ListTunnels (Maybe Text)

The name of the IoT thing associated with the destination device.

listTunnels_maxResults :: Lens' ListTunnels (Maybe Natural)

The maximum number of results to return at once.

Destructuring the Response

data ListTunnelsResponse

See: newListTunnelsResponse smart constructor.

Constructors

ListTunnelsResponse'

Fields

nextToken :: Maybe Text

A token to used to retrieve the next set of results.

tunnelSummaries :: Maybe [TunnelSummary]

A short description of the tunnels in an AWS account.

httpStatus :: Int

The response's http status code.

newListTunnelsResponse

Arguments

:: Int
$sel:httpStatus:ListTunnelsResponse'

-> ListTunnelsResponse

Create a value of ListTunnelsResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:nextToken:ListTunnels', listTunnelsResponse_nextToken - A token to used to retrieve the next set of results.

$sel:tunnelSummaries:ListTunnelsResponse', listTunnelsResponse_tunnelSummaries - A short description of the tunnels in an AWS account.

$sel:httpStatus:ListTunnelsResponse', listTunnelsResponse_httpStatus - The response's http status code.

Response Lenses

listTunnelsResponse_nextToken :: Lens' ListTunnelsResponse (Maybe Text)

A token to used to retrieve the next set of results.

listTunnelsResponse_tunnelSummaries :: Lens' ListTunnelsResponse (Maybe [TunnelSummary])

A short description of the tunnels in an AWS account.

listTunnelsResponse_httpStatus :: Lens' ListTunnelsResponse Int

The response's http status code.