Amazonka.IoTSecureTunneling.Types

Description

Service Configuration

defaultService :: Service

API version 2018-10-05 of the Amazon IoT Secure Tunneling SDK configuration.

Errors

_ResourceNotFoundException :: AsError a => Getting (First ServiceError) a ServiceError

Thrown when an operation is attempted on a resource that does not exist.

_LimitExceededException :: AsError a => Getting (First ServiceError) a ServiceError

Thrown when a tunnel limit is exceeded.

ConnectionStatus

newtype ConnectionStatus

Constructors

ConnectionStatus'

Fields

fromConnectionStatus :: Text

Bundled Patterns

pattern ConnectionStatus_CONNECTED :: ConnectionStatus

pattern ConnectionStatus_DISCONNECTED :: ConnectionStatus

TunnelStatus

newtype TunnelStatus

Constructors

TunnelStatus'

Fields

fromTunnelStatus :: Text

Bundled Patterns

pattern TunnelStatus_CLOSED :: TunnelStatus

pattern TunnelStatus_OPEN :: TunnelStatus

ConnectionState

data ConnectionState

The state of a connection.

See: newConnectionState smart constructor.

Constructors

ConnectionState'

Fields

status :: Maybe ConnectionStatus

The connection status of the tunnel. Valid values are CONNECTED and DISCONNECTED.

lastUpdatedAt :: Maybe POSIX

The last time the connection status was updated.

newConnectionState :: ConnectionState

Create a value of ConnectionState with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:status:ConnectionState', connectionState_status - The connection status of the tunnel. Valid values are CONNECTED and DISCONNECTED.

$sel:lastUpdatedAt:ConnectionState', connectionState_lastUpdatedAt - The last time the connection status was updated.

connectionState_status :: Lens' ConnectionState (Maybe ConnectionStatus)

The connection status of the tunnel. Valid values are CONNECTED and DISCONNECTED.

connectionState_lastUpdatedAt :: Lens' ConnectionState (Maybe UTCTime)

The last time the connection status was updated.

DestinationConfig

data DestinationConfig

The destination configuration.

See: newDestinationConfig smart constructor.

Constructors

DestinationConfig'

Fields

thingName :: Maybe Text

The name of the IoT thing to which you want to connect.

services :: NonEmpty Text

A list of service names that identity the target application. The AWS IoT client running on the destination device reads this value and uses it to look up a port or an IP address and a port. The AWS IoT client instantiates the local proxy which uses this information to connect to the destination application.

newDestinationConfig

Arguments

:: NonEmpty Text
$sel:services:DestinationConfig'

-> DestinationConfig

Create a value of DestinationConfig with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:thingName:DestinationConfig', destinationConfig_thingName - The name of the IoT thing to which you want to connect.

$sel:services:DestinationConfig', destinationConfig_services - A list of service names that identity the target application. The AWS IoT client running on the destination device reads this value and uses it to look up a port or an IP address and a port. The AWS IoT client instantiates the local proxy which uses this information to connect to the destination application.

destinationConfig_thingName :: Lens' DestinationConfig (Maybe Text)

The name of the IoT thing to which you want to connect.

destinationConfig_services :: Lens' DestinationConfig (NonEmpty Text)

A list of service names that identity the target application. The AWS IoT client running on the destination device reads this value and uses it to look up a port or an IP address and a port. The AWS IoT client instantiates the local proxy which uses this information to connect to the destination application.

Tag

data Tag

An arbitary key/value pair used to add searchable metadata to secure tunnel resources.

See: newTag smart constructor.

Constructors

Tag'

Fields

key :: Text

The key of the tag.

value :: Text

The value of the tag.

newTag

Arguments

:: Text
$sel:key:Tag'

-> Text
$sel:value:Tag'

-> Tag

Create a value of Tag with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:key:Tag', tag_key - The key of the tag.

$sel:value:Tag', tag_value - The value of the tag.

tag_key :: Lens' Tag Text

The key of the tag.

tag_value :: Lens' Tag Text

The value of the tag.

TimeoutConfig

data TimeoutConfig

Tunnel timeout configuration.

See: newTimeoutConfig smart constructor.

Constructors

TimeoutConfig'

Fields

maxLifetimeTimeoutMinutes :: Maybe Natural

The maximum amount of time (in minutes) a tunnel can remain open. If not specified, maxLifetimeTimeoutMinutes defaults to 720 minutes. Valid values are from 1 minute to 12 hours (720 minutes)

newTimeoutConfig :: TimeoutConfig

Create a value of TimeoutConfig with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:maxLifetimeTimeoutMinutes:TimeoutConfig', timeoutConfig_maxLifetimeTimeoutMinutes - The maximum amount of time (in minutes) a tunnel can remain open. If not specified, maxLifetimeTimeoutMinutes defaults to 720 minutes. Valid values are from 1 minute to 12 hours (720 minutes)

timeoutConfig_maxLifetimeTimeoutMinutes :: Lens' TimeoutConfig (Maybe Natural)

The maximum amount of time (in minutes) a tunnel can remain open. If not specified, maxLifetimeTimeoutMinutes defaults to 720 minutes. Valid values are from 1 minute to 12 hours (720 minutes)

Tunnel

data Tunnel

A connection between a source computer and a destination device.

See: newTunnel smart constructor.

Constructors

Tunnel'

Fields

status :: Maybe TunnelStatus

The status of a tunnel. Valid values are: Open and Closed.

lastUpdatedAt :: Maybe POSIX

The last time the tunnel was updated.

createdAt :: Maybe POSIX

The time when the tunnel was created.

tunnelArn :: Maybe Text

The Amazon Resource Name (ARN) of a tunnel. The tunnel ARN format is arn:aws:tunnel:<region>:<account-id>:tunnel/<tunnel-id>

sourceConnectionState :: Maybe ConnectionState

The connection state of the source application.

destinationConnectionState :: Maybe ConnectionState

The connection state of the destination application.

tunnelId :: Maybe Text

A unique alpha-numeric ID that identifies a tunnel.

destinationConfig :: Maybe DestinationConfig

The destination configuration that specifies the thing name of the destination device and a service name that the local proxy uses to connect to the destination application.

description :: Maybe Text

A description of the tunnel.

timeoutConfig :: Maybe TimeoutConfig

Timeout configuration for the tunnel.

tags :: Maybe (NonEmpty Tag)

A list of tag metadata associated with the secure tunnel.

newTunnel :: Tunnel

Create a value of Tunnel with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:status:Tunnel', tunnel_status - The status of a tunnel. Valid values are: Open and Closed.

$sel:lastUpdatedAt:Tunnel', tunnel_lastUpdatedAt - The last time the tunnel was updated.

$sel:createdAt:Tunnel', tunnel_createdAt - The time when the tunnel was created.

$sel:tunnelArn:Tunnel', tunnel_tunnelArn - The Amazon Resource Name (ARN) of a tunnel. The tunnel ARN format is arn:aws:tunnel:<region>:<account-id>:tunnel/<tunnel-id>

$sel:sourceConnectionState:Tunnel', tunnel_sourceConnectionState - The connection state of the source application.

$sel:destinationConnectionState:Tunnel', tunnel_destinationConnectionState - The connection state of the destination application.

$sel:tunnelId:Tunnel', tunnel_tunnelId - A unique alpha-numeric ID that identifies a tunnel.

$sel:destinationConfig:Tunnel', tunnel_destinationConfig - The destination configuration that specifies the thing name of the destination device and a service name that the local proxy uses to connect to the destination application.

$sel:description:Tunnel', tunnel_description - A description of the tunnel.

$sel:timeoutConfig:Tunnel', tunnel_timeoutConfig - Timeout configuration for the tunnel.

$sel:tags:Tunnel', tunnel_tags - A list of tag metadata associated with the secure tunnel.

tunnel_status :: Lens' Tunnel (Maybe TunnelStatus)

The status of a tunnel. Valid values are: Open and Closed.

tunnel_lastUpdatedAt :: Lens' Tunnel (Maybe UTCTime)

The last time the tunnel was updated.

tunnel_createdAt :: Lens' Tunnel (Maybe UTCTime)

The time when the tunnel was created.

tunnel_tunnelArn :: Lens' Tunnel (Maybe Text)

The Amazon Resource Name (ARN) of a tunnel. The tunnel ARN format is arn:aws:tunnel:<region>:<account-id>:tunnel/<tunnel-id>

tunnel_sourceConnectionState :: Lens' Tunnel (Maybe ConnectionState)

The connection state of the source application.

tunnel_destinationConnectionState :: Lens' Tunnel (Maybe ConnectionState)

The connection state of the destination application.

tunnel_tunnelId :: Lens' Tunnel (Maybe Text)

A unique alpha-numeric ID that identifies a tunnel.

tunnel_destinationConfig :: Lens' Tunnel (Maybe DestinationConfig)

The destination configuration that specifies the thing name of the destination device and a service name that the local proxy uses to connect to the destination application.

tunnel_description :: Lens' Tunnel (Maybe Text)

A description of the tunnel.

tunnel_timeoutConfig :: Lens' Tunnel (Maybe TimeoutConfig)

Timeout configuration for the tunnel.

tunnel_tags :: Lens' Tunnel (Maybe (NonEmpty Tag))

A list of tag metadata associated with the secure tunnel.

TunnelSummary

data TunnelSummary

Information about the tunnel.

See: newTunnelSummary smart constructor.

Constructors

TunnelSummary'

Fields

status :: Maybe TunnelStatus

The status of a tunnel. Valid values are: Open and Closed.

lastUpdatedAt :: Maybe POSIX

The time the tunnel was last updated.

createdAt :: Maybe POSIX

The time the tunnel was created.

tunnelArn :: Maybe Text

The Amazon Resource Name of the tunnel. The tunnel ARN format is arn:aws:tunnel:<region>:<account-id>:tunnel/<tunnel-id>

tunnelId :: Maybe Text

The unique alpha-numeric identifier for the tunnel.

description :: Maybe Text

A description of the tunnel.

newTunnelSummary :: TunnelSummary

Create a value of TunnelSummary with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:status:TunnelSummary', tunnelSummary_status - The status of a tunnel. Valid values are: Open and Closed.

$sel:lastUpdatedAt:TunnelSummary', tunnelSummary_lastUpdatedAt - The time the tunnel was last updated.

$sel:createdAt:TunnelSummary', tunnelSummary_createdAt - The time the tunnel was created.

$sel:tunnelArn:TunnelSummary', tunnelSummary_tunnelArn - The Amazon Resource Name of the tunnel. The tunnel ARN format is arn:aws:tunnel:<region>:<account-id>:tunnel/<tunnel-id>

$sel:tunnelId:TunnelSummary', tunnelSummary_tunnelId - The unique alpha-numeric identifier for the tunnel.

$sel:description:TunnelSummary', tunnelSummary_description - A description of the tunnel.

tunnelSummary_status :: Lens' TunnelSummary (Maybe TunnelStatus)

The status of a tunnel. Valid values are: Open and Closed.

tunnelSummary_lastUpdatedAt :: Lens' TunnelSummary (Maybe UTCTime)

The time the tunnel was last updated.

tunnelSummary_createdAt :: Lens' TunnelSummary (Maybe UTCTime)

The time the tunnel was created.

tunnelSummary_tunnelArn :: Lens' TunnelSummary (Maybe Text)

The Amazon Resource Name of the tunnel. The tunnel ARN format is arn:aws:tunnel:<region>:<account-id>:tunnel/<tunnel-id>

tunnelSummary_tunnelId :: Lens' TunnelSummary (Maybe Text)

The unique alpha-numeric identifier for the tunnel.

tunnelSummary_description :: Lens' TunnelSummary (Maybe Text)

A description of the tunnel.