Amazonka.IoTSiteWise.BatchAssociateProjectAssets

Description

Associates a group (batch) of assets with an IoT SiteWise Monitor project.

Creating a Request

data BatchAssociateProjectAssets

See: newBatchAssociateProjectAssets smart constructor.

Constructors

BatchAssociateProjectAssets'

Fields

clientToken :: Maybe Text

A unique case-sensitive identifier that you can provide to ensure the idempotency of the request. Don't reuse this client token if a new idempotent request is required.

projectId :: Text

The ID of the project to which to associate the assets.

assetIds :: NonEmpty Text

The IDs of the assets to be associated to the project.

newBatchAssociateProjectAssets

Arguments

:: Text
$sel:projectId:BatchAssociateProjectAssets'

-> NonEmpty Text
$sel:assetIds:BatchAssociateProjectAssets'

-> BatchAssociateProjectAssets

Create a value of BatchAssociateProjectAssets with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:clientToken:BatchAssociateProjectAssets', batchAssociateProjectAssets_clientToken - A unique case-sensitive identifier that you can provide to ensure the idempotency of the request. Don't reuse this client token if a new idempotent request is required.

$sel:projectId:BatchAssociateProjectAssets', batchAssociateProjectAssets_projectId - The ID of the project to which to associate the assets.

$sel:assetIds:BatchAssociateProjectAssets', batchAssociateProjectAssets_assetIds - The IDs of the assets to be associated to the project.

Request Lenses

batchAssociateProjectAssets_clientToken :: Lens' BatchAssociateProjectAssets (Maybe Text)

A unique case-sensitive identifier that you can provide to ensure the idempotency of the request. Don't reuse this client token if a new idempotent request is required.

batchAssociateProjectAssets_projectId :: Lens' BatchAssociateProjectAssets Text

The ID of the project to which to associate the assets.

batchAssociateProjectAssets_assetIds :: Lens' BatchAssociateProjectAssets (NonEmpty Text)

The IDs of the assets to be associated to the project.

Destructuring the Response

data BatchAssociateProjectAssetsResponse

See: newBatchAssociateProjectAssetsResponse smart constructor.

Constructors

BatchAssociateProjectAssetsResponse'

Fields

errors :: Maybe [AssetErrorDetails]

A list of associated error information, if any.

httpStatus :: Int

The response's http status code.

newBatchAssociateProjectAssetsResponse

Arguments

:: Int
$sel:httpStatus:BatchAssociateProjectAssetsResponse'

-> BatchAssociateProjectAssetsResponse

Create a value of BatchAssociateProjectAssetsResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:errors:BatchAssociateProjectAssetsResponse', batchAssociateProjectAssetsResponse_errors - A list of associated error information, if any.

$sel:httpStatus:BatchAssociateProjectAssetsResponse', batchAssociateProjectAssetsResponse_httpStatus - The response's http status code.

Response Lenses

batchAssociateProjectAssetsResponse_errors :: Lens' BatchAssociateProjectAssetsResponse (Maybe [AssetErrorDetails])

A list of associated error information, if any.

batchAssociateProjectAssetsResponse_httpStatus :: Lens' BatchAssociateProjectAssetsResponse Int

The response's http status code.