Amazonka.IoTSiteWise.BatchDisassociateProjectAssets

Description

Disassociates a group (batch) of assets from an IoT SiteWise Monitor project.

Creating a Request

data BatchDisassociateProjectAssets

See: newBatchDisassociateProjectAssets smart constructor.

Constructors

BatchDisassociateProjectAssets'

Fields

clientToken :: Maybe Text

A unique case-sensitive identifier that you can provide to ensure the idempotency of the request. Don't reuse this client token if a new idempotent request is required.

projectId :: Text

The ID of the project from which to disassociate the assets.

assetIds :: NonEmpty Text

The IDs of the assets to be disassociated from the project.

newBatchDisassociateProjectAssets

Arguments

:: Text
$sel:projectId:BatchDisassociateProjectAssets'

-> NonEmpty Text
$sel:assetIds:BatchDisassociateProjectAssets'

-> BatchDisassociateProjectAssets

Create a value of BatchDisassociateProjectAssets with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:clientToken:BatchDisassociateProjectAssets', batchDisassociateProjectAssets_clientToken - A unique case-sensitive identifier that you can provide to ensure the idempotency of the request. Don't reuse this client token if a new idempotent request is required.

$sel:projectId:BatchDisassociateProjectAssets', batchDisassociateProjectAssets_projectId - The ID of the project from which to disassociate the assets.

$sel:assetIds:BatchDisassociateProjectAssets', batchDisassociateProjectAssets_assetIds - The IDs of the assets to be disassociated from the project.

Request Lenses

batchDisassociateProjectAssets_clientToken :: Lens' BatchDisassociateProjectAssets (Maybe Text)

A unique case-sensitive identifier that you can provide to ensure the idempotency of the request. Don't reuse this client token if a new idempotent request is required.

batchDisassociateProjectAssets_projectId :: Lens' BatchDisassociateProjectAssets Text

The ID of the project from which to disassociate the assets.

batchDisassociateProjectAssets_assetIds :: Lens' BatchDisassociateProjectAssets (NonEmpty Text)

The IDs of the assets to be disassociated from the project.

Destructuring the Response

data BatchDisassociateProjectAssetsResponse

See: newBatchDisassociateProjectAssetsResponse smart constructor.

Constructors

BatchDisassociateProjectAssetsResponse'

Fields

errors :: Maybe [AssetErrorDetails]

A list of associated error information, if any.

httpStatus :: Int

The response's http status code.

newBatchDisassociateProjectAssetsResponse

Arguments

:: Int
$sel:httpStatus:BatchDisassociateProjectAssetsResponse'

-> BatchDisassociateProjectAssetsResponse

Create a value of BatchDisassociateProjectAssetsResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:errors:BatchDisassociateProjectAssetsResponse', batchDisassociateProjectAssetsResponse_errors - A list of associated error information, if any.

$sel:httpStatus:BatchDisassociateProjectAssetsResponse', batchDisassociateProjectAssetsResponse_httpStatus - The response's http status code.

Response Lenses

batchDisassociateProjectAssetsResponse_errors :: Lens' BatchDisassociateProjectAssetsResponse (Maybe [AssetErrorDetails])

A list of associated error information, if any.

batchDisassociateProjectAssetsResponse_httpStatus :: Lens' BatchDisassociateProjectAssetsResponse Int

The response's http status code.