Amazonka.IoTSiteWise.BatchPutAssetPropertyValue

Description

Sends a list of asset property values to IoT SiteWise. Each value is a timestamp-quality-value (TQV) data point. For more information, see Ingesting data using the API in the IoT SiteWise User Guide.

To identify an asset property, you must specify one of the following:

The assetId and propertyId of an asset property.

A propertyAlias, which is a data stream alias (for example, /company/windfarm/3/turbine/7/temperature). To define an asset property's alias, see UpdateAssetProperty.

With respect to Unix epoch time, IoT SiteWise accepts only TQVs that have a timestamp of no more than 7 days in the past and no more than 10 minutes in the future. IoT SiteWise rejects timestamps outside of the inclusive range of [-7 days, +10 minutes] and returns a TimestampOutOfRangeException error.

For each asset property, IoT SiteWise overwrites TQVs with duplicate timestamps unless the newer TQV has a different quality. For example, if you store a TQV {T1, GOOD, V1}, then storing {T1, GOOD, V2} replaces the existing TQV.

IoT SiteWise authorizes access to each BatchPutAssetPropertyValue entry individually. For more information, see BatchPutAssetPropertyValue authorization in the IoT SiteWise User Guide.

Creating a Request

data BatchPutAssetPropertyValue

See: newBatchPutAssetPropertyValue smart constructor.

Constructors

BatchPutAssetPropertyValue'

Fields

entries :: [PutAssetPropertyValueEntry]

The list of asset property value entries for the batch put request. You can specify up to 10 entries per request.

newBatchPutAssetPropertyValue :: BatchPutAssetPropertyValue

Create a value of BatchPutAssetPropertyValue with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:entries:BatchPutAssetPropertyValue', batchPutAssetPropertyValue_entries - The list of asset property value entries for the batch put request. You can specify up to 10 entries per request.

Request Lenses

batchPutAssetPropertyValue_entries :: Lens' BatchPutAssetPropertyValue [PutAssetPropertyValueEntry]

The list of asset property value entries for the batch put request. You can specify up to 10 entries per request.

Destructuring the Response

data BatchPutAssetPropertyValueResponse

See: newBatchPutAssetPropertyValueResponse smart constructor.

Constructors

BatchPutAssetPropertyValueResponse'

Fields

httpStatus :: Int

The response's http status code.

errorEntries :: [BatchPutAssetPropertyErrorEntry]

A list of the errors (if any) associated with the batch put request. Each error entry contains the entryId of the entry that failed.

newBatchPutAssetPropertyValueResponse

Arguments

:: Int
$sel:httpStatus:BatchPutAssetPropertyValueResponse'

-> BatchPutAssetPropertyValueResponse

Create a value of BatchPutAssetPropertyValueResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:httpStatus:BatchPutAssetPropertyValueResponse', batchPutAssetPropertyValueResponse_httpStatus - The response's http status code.

$sel:errorEntries:BatchPutAssetPropertyValueResponse', batchPutAssetPropertyValueResponse_errorEntries - A list of the errors (if any) associated with the batch put request. Each error entry contains the entryId of the entry that failed.

Response Lenses

batchPutAssetPropertyValueResponse_httpStatus :: Lens' BatchPutAssetPropertyValueResponse Int

The response's http status code.

batchPutAssetPropertyValueResponse_errorEntries :: Lens' BatchPutAssetPropertyValueResponse [BatchPutAssetPropertyErrorEntry]

A list of the errors (if any) associated with the batch put request. Each error entry contains the entryId of the entry that failed.