Amazonka.IoTSiteWise.CreateDashboard

Description

Creates a dashboard in an IoT SiteWise Monitor project.

Creating a Request

data CreateDashboard

See: newCreateDashboard smart constructor.

Constructors

CreateDashboard'

Fields

clientToken :: Maybe Text

A unique case-sensitive identifier that you can provide to ensure the idempotency of the request. Don't reuse this client token if a new idempotent request is required.

dashboardDescription :: Maybe Text

A description for the dashboard.

tags :: Maybe (HashMap Text Text)

A list of key-value pairs that contain metadata for the dashboard. For more information, see Tagging your IoT SiteWise resources in the IoT SiteWise User Guide.

projectId :: Text

The ID of the project in which to create the dashboard.

dashboardName :: Text

A friendly name for the dashboard.

dashboardDefinition :: Text

The dashboard definition specified in a JSON literal. For detailed information, see Creating dashboards (CLI) in the IoT SiteWise User Guide.

newCreateDashboard

Arguments

:: Text
$sel:projectId:CreateDashboard'

-> Text
$sel:dashboardName:CreateDashboard'

-> Text
$sel:dashboardDefinition:CreateDashboard'

-> CreateDashboard

Create a value of CreateDashboard with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:clientToken:CreateDashboard', createDashboard_clientToken - A unique case-sensitive identifier that you can provide to ensure the idempotency of the request. Don't reuse this client token if a new idempotent request is required.

$sel:dashboardDescription:CreateDashboard', createDashboard_dashboardDescription - A description for the dashboard.

$sel:tags:CreateDashboard', createDashboard_tags - A list of key-value pairs that contain metadata for the dashboard. For more information, see Tagging your IoT SiteWise resources in the IoT SiteWise User Guide.

$sel:projectId:CreateDashboard', createDashboard_projectId - The ID of the project in which to create the dashboard.

$sel:dashboardName:CreateDashboard', createDashboard_dashboardName - A friendly name for the dashboard.

$sel:dashboardDefinition:CreateDashboard', createDashboard_dashboardDefinition - The dashboard definition specified in a JSON literal. For detailed information, see Creating dashboards (CLI) in the IoT SiteWise User Guide.

Request Lenses

createDashboard_clientToken :: Lens' CreateDashboard (Maybe Text)

A unique case-sensitive identifier that you can provide to ensure the idempotency of the request. Don't reuse this client token if a new idempotent request is required.

createDashboard_dashboardDescription :: Lens' CreateDashboard (Maybe Text)

A description for the dashboard.

createDashboard_tags :: Lens' CreateDashboard (Maybe (HashMap Text Text))

A list of key-value pairs that contain metadata for the dashboard. For more information, see Tagging your IoT SiteWise resources in the IoT SiteWise User Guide.

createDashboard_projectId :: Lens' CreateDashboard Text

The ID of the project in which to create the dashboard.

createDashboard_dashboardName :: Lens' CreateDashboard Text

A friendly name for the dashboard.

createDashboard_dashboardDefinition :: Lens' CreateDashboard Text

The dashboard definition specified in a JSON literal. For detailed information, see Creating dashboards (CLI) in the IoT SiteWise User Guide.

Destructuring the Response

data CreateDashboardResponse

See: newCreateDashboardResponse smart constructor.

Constructors

CreateDashboardResponse'

Fields

httpStatus :: Int

The response's http status code.

dashboardId :: Text

The ID of the dashboard.

dashboardArn :: Text

The ARN of the dashboard, which has the following format.

arn:${Partition}:iotsitewise:${Region}:${Account}:dashboard/${DashboardId}

newCreateDashboardResponse

Arguments

:: Int
$sel:httpStatus:CreateDashboardResponse'

-> Text
$sel:dashboardId:CreateDashboardResponse'

-> Text
$sel:dashboardArn:CreateDashboardResponse'

-> CreateDashboardResponse

Create a value of CreateDashboardResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:httpStatus:CreateDashboardResponse', createDashboardResponse_httpStatus - The response's http status code.

$sel:dashboardId:CreateDashboardResponse', createDashboardResponse_dashboardId - The ID of the dashboard.

$sel:dashboardArn:CreateDashboardResponse', createDashboardResponse_dashboardArn - The ARN of the dashboard, which has the following format.

arn:${Partition}:iotsitewise:${Region}:${Account}:dashboard/${DashboardId}

Response Lenses

createDashboardResponse_httpStatus :: Lens' CreateDashboardResponse Int

The response's http status code.

createDashboardResponse_dashboardId :: Lens' CreateDashboardResponse Text

The ID of the dashboard.

createDashboardResponse_dashboardArn :: Lens' CreateDashboardResponse Text

The ARN of the dashboard, which has the following format.

arn:${Partition}:iotsitewise:${Region}:${Account}:dashboard/${DashboardId}