Amazonka.IoTSiteWise.DeletePortal

Description

Deletes a portal from IoT SiteWise Monitor.

Creating a Request

data DeletePortal

See: newDeletePortal smart constructor.

Constructors

DeletePortal'

Fields

clientToken :: Maybe Text

A unique case-sensitive identifier that you can provide to ensure the idempotency of the request. Don't reuse this client token if a new idempotent request is required.

portalId :: Text

The ID of the portal to delete.

newDeletePortal

Arguments

:: Text
$sel:portalId:DeletePortal'

-> DeletePortal

Create a value of DeletePortal with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:clientToken:DeletePortal', deletePortal_clientToken - A unique case-sensitive identifier that you can provide to ensure the idempotency of the request. Don't reuse this client token if a new idempotent request is required.

$sel:portalId:DeletePortal', deletePortal_portalId - The ID of the portal to delete.

Request Lenses

deletePortal_clientToken :: Lens' DeletePortal (Maybe Text)

A unique case-sensitive identifier that you can provide to ensure the idempotency of the request. Don't reuse this client token if a new idempotent request is required.

deletePortal_portalId :: Lens' DeletePortal Text

The ID of the portal to delete.

Destructuring the Response

data DeletePortalResponse

See: newDeletePortalResponse smart constructor.

Constructors

DeletePortalResponse'

Fields

httpStatus :: Int

The response's http status code.

portalStatus :: PortalStatus

The status of the portal, which contains a state (DELETING after successfully calling this operation) and any error message.

newDeletePortalResponse

Arguments

:: Int
$sel:httpStatus:DeletePortalResponse'

-> PortalStatus
$sel:portalStatus:DeletePortalResponse'

-> DeletePortalResponse

Create a value of DeletePortalResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:httpStatus:DeletePortalResponse', deletePortalResponse_httpStatus - The response's http status code.

$sel:portalStatus:DeletePortalResponse', deletePortalResponse_portalStatus - The status of the portal, which contains a state (DELETING after successfully calling this operation) and any error message.

Response Lenses

deletePortalResponse_httpStatus :: Lens' DeletePortalResponse Int

The response's http status code.

deletePortalResponse_portalStatus :: Lens' DeletePortalResponse PortalStatus

The status of the portal, which contains a state (DELETING after successfully calling this operation) and any error message.