Amazonka.IoTSiteWise.DescribeAssetProperty

Description

Retrieves information about an asset property.

When you call this operation for an attribute property, this response includes the default attribute value that you define in the asset model. If you update the default value in the model, this operation's response includes the new default value.

This operation doesn't return the value of the asset property. To get the value of an asset property, use GetAssetPropertyValue.

Creating a Request

data DescribeAssetProperty

See: newDescribeAssetProperty smart constructor.

Constructors

DescribeAssetProperty'

Fields

assetId :: Text

The ID of the asset.

propertyId :: Text

The ID of the asset property.

newDescribeAssetProperty

Arguments

:: Text
$sel:assetId:DescribeAssetProperty'

-> Text
$sel:propertyId:DescribeAssetProperty'

-> DescribeAssetProperty

Create a value of DescribeAssetProperty with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:assetId:DescribeAssetProperty', describeAssetProperty_assetId - The ID of the asset.

$sel:propertyId:DescribeAssetProperty', describeAssetProperty_propertyId - The ID of the asset property.

Request Lenses

describeAssetProperty_assetId :: Lens' DescribeAssetProperty Text

The ID of the asset.

describeAssetProperty_propertyId :: Lens' DescribeAssetProperty Text

The ID of the asset property.

Destructuring the Response

data DescribeAssetPropertyResponse

See: newDescribeAssetPropertyResponse smart constructor.

Constructors

DescribeAssetPropertyResponse'

Fields

assetProperty :: Maybe Property

The asset property's definition, alias, and notification state.

This response includes this object for normal asset properties. If you describe an asset property in a composite model, this response includes the asset property information in compositeModel.

compositeModel :: Maybe CompositeModelProperty

The composite asset model that declares this asset property, if this asset property exists in a composite model.

httpStatus :: Int

The response's http status code.

assetId :: Text

The ID of the asset.

assetName :: Text

The name of the asset.

assetModelId :: Text

The ID of the asset model.

newDescribeAssetPropertyResponse

Arguments

:: Int
$sel:httpStatus:DescribeAssetPropertyResponse'

-> Text
$sel:assetId:DescribeAssetProperty'

-> Text
$sel:assetName:DescribeAssetPropertyResponse'

-> Text
$sel:assetModelId:DescribeAssetPropertyResponse'

-> DescribeAssetPropertyResponse

Create a value of DescribeAssetPropertyResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:assetProperty:DescribeAssetPropertyResponse', describeAssetPropertyResponse_assetProperty - The asset property's definition, alias, and notification state.

This response includes this object for normal asset properties. If you describe an asset property in a composite model, this response includes the asset property information in compositeModel.

$sel:compositeModel:DescribeAssetPropertyResponse', describeAssetPropertyResponse_compositeModel - The composite asset model that declares this asset property, if this asset property exists in a composite model.

$sel:httpStatus:DescribeAssetPropertyResponse', describeAssetPropertyResponse_httpStatus - The response's http status code.

$sel:assetId:DescribeAssetProperty', describeAssetPropertyResponse_assetId - The ID of the asset.

$sel:assetName:DescribeAssetPropertyResponse', describeAssetPropertyResponse_assetName - The name of the asset.

$sel:assetModelId:DescribeAssetPropertyResponse', describeAssetPropertyResponse_assetModelId - The ID of the asset model.

Response Lenses

describeAssetPropertyResponse_assetProperty :: Lens' DescribeAssetPropertyResponse (Maybe Property)

The asset property's definition, alias, and notification state.

This response includes this object for normal asset properties. If you describe an asset property in a composite model, this response includes the asset property information in compositeModel.

describeAssetPropertyResponse_compositeModel :: Lens' DescribeAssetPropertyResponse (Maybe CompositeModelProperty)

The composite asset model that declares this asset property, if this asset property exists in a composite model.

describeAssetPropertyResponse_httpStatus :: Lens' DescribeAssetPropertyResponse Int

The response's http status code.

describeAssetPropertyResponse_assetId :: Lens' DescribeAssetPropertyResponse Text

The ID of the asset.

describeAssetPropertyResponse_assetName :: Lens' DescribeAssetPropertyResponse Text

The name of the asset.

describeAssetPropertyResponse_assetModelId :: Lens' DescribeAssetPropertyResponse Text

The ID of the asset model.