Amazonka.IoTSiteWise.DescribeDefaultEncryptionConfiguration

Description

Retrieves information about the default encryption configuration for the Amazon Web Services account in the default or specified Region. For more information, see Key management in the IoT SiteWise User Guide.

Creating a Request

data DescribeDefaultEncryptionConfiguration

See: newDescribeDefaultEncryptionConfiguration smart constructor.

Constructors

DescribeDefaultEncryptionConfiguration'

newDescribeDefaultEncryptionConfiguration :: DescribeDefaultEncryptionConfiguration

Create a value of DescribeDefaultEncryptionConfiguration with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

Destructuring the Response

data DescribeDefaultEncryptionConfigurationResponse

See: newDescribeDefaultEncryptionConfigurationResponse smart constructor.

Constructors

DescribeDefaultEncryptionConfigurationResponse'

Fields

kmsKeyArn :: Maybe Text

The key ARN of the customer managed customer master key (CMK) used for KMS encryption if you use KMS_BASED_ENCRYPTION.

httpStatus :: Int

The response's http status code.

encryptionType :: EncryptionType

The type of encryption used for the encryption configuration.

configurationStatus :: ConfigurationStatus

The status of the account configuration. This contains the ConfigurationState. If there's an error, it also contains the ErrorDetails.

newDescribeDefaultEncryptionConfigurationResponse

Arguments

:: Int
$sel:httpStatus:DescribeDefaultEncryptionConfigurationResponse'

-> EncryptionType
$sel:encryptionType:DescribeDefaultEncryptionConfigurationResponse'

-> ConfigurationStatus
$sel:configurationStatus:DescribeDefaultEncryptionConfigurationResponse'

-> DescribeDefaultEncryptionConfigurationResponse

Create a value of DescribeDefaultEncryptionConfigurationResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:kmsKeyArn:DescribeDefaultEncryptionConfigurationResponse', describeDefaultEncryptionConfigurationResponse_kmsKeyArn - The key ARN of the customer managed customer master key (CMK) used for KMS encryption if you use KMS_BASED_ENCRYPTION.

$sel:httpStatus:DescribeDefaultEncryptionConfigurationResponse', describeDefaultEncryptionConfigurationResponse_httpStatus - The response's http status code.

$sel:encryptionType:DescribeDefaultEncryptionConfigurationResponse', describeDefaultEncryptionConfigurationResponse_encryptionType - The type of encryption used for the encryption configuration.

$sel:configurationStatus:DescribeDefaultEncryptionConfigurationResponse', describeDefaultEncryptionConfigurationResponse_configurationStatus - The status of the account configuration. This contains the ConfigurationState. If there's an error, it also contains the ErrorDetails.

Response Lenses

describeDefaultEncryptionConfigurationResponse_kmsKeyArn :: Lens' DescribeDefaultEncryptionConfigurationResponse (Maybe Text)

The key ARN of the customer managed customer master key (CMK) used for KMS encryption if you use KMS_BASED_ENCRYPTION.

describeDefaultEncryptionConfigurationResponse_httpStatus :: Lens' DescribeDefaultEncryptionConfigurationResponse Int

The response's http status code.

describeDefaultEncryptionConfigurationResponse_encryptionType :: Lens' DescribeDefaultEncryptionConfigurationResponse EncryptionType

The type of encryption used for the encryption configuration.

describeDefaultEncryptionConfigurationResponse_configurationStatus :: Lens' DescribeDefaultEncryptionConfigurationResponse ConfigurationStatus

The status of the account configuration. This contains the ConfigurationState. If there's an error, it also contains the ErrorDetails.