Amazonka.IoTSiteWise.GetAssetPropertyAggregates

Description

Gets aggregated values for an asset property. For more information, see Querying aggregates in the IoT SiteWise User Guide.

To identify an asset property, you must specify one of the following:

The assetId and propertyId of an asset property.

A propertyAlias, which is a data stream alias (for example, /company/windfarm/3/turbine/7/temperature). To define an asset property's alias, see UpdateAssetProperty.

This operation returns paginated results.

Creating a Request

data GetAssetPropertyAggregates

See: newGetAssetPropertyAggregates smart constructor.

Constructors

GetAssetPropertyAggregates'

Fields

qualities :: Maybe (NonEmpty Quality)

The quality by which to filter asset data.

timeOrdering :: Maybe TimeOrdering

The chronological sorting order of the requested information.

Default: ASCENDING

propertyAlias :: Maybe Text

The alias that identifies the property, such as an OPC-UA server data stream path (for example, /company/windfarm/3/turbine/7/temperature). For more information, see Mapping industrial data streams to asset properties in the IoT SiteWise User Guide.

nextToken :: Maybe Text

The token to be used for the next set of paginated results.

propertyId :: Maybe Text

The ID of the asset property.

assetId :: Maybe Text

The ID of the asset.

maxResults :: Maybe Natural

The maximum number of results to return for each paginated request.

Default: 100

aggregateTypes :: NonEmpty AggregateType

The data aggregating function.

resolution :: Text

The time interval over which to aggregate data.

startDate :: POSIX

The exclusive start of the range from which to query historical data, expressed in seconds in Unix epoch time.

endDate :: POSIX

The inclusive end of the range from which to query historical data, expressed in seconds in Unix epoch time.

newGetAssetPropertyAggregates

Arguments

:: NonEmpty AggregateType
$sel:aggregateTypes:GetAssetPropertyAggregates'

-> Text
$sel:resolution:GetAssetPropertyAggregates'

-> UTCTime
$sel:startDate:GetAssetPropertyAggregates'

-> UTCTime
$sel:endDate:GetAssetPropertyAggregates'

-> GetAssetPropertyAggregates

Create a value of GetAssetPropertyAggregates with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:qualities:GetAssetPropertyAggregates', getAssetPropertyAggregates_qualities - The quality by which to filter asset data.

$sel:timeOrdering:GetAssetPropertyAggregates', getAssetPropertyAggregates_timeOrdering - The chronological sorting order of the requested information.

Default: ASCENDING

$sel:propertyAlias:GetAssetPropertyAggregates', getAssetPropertyAggregates_propertyAlias - The alias that identifies the property, such as an OPC-UA server data stream path (for example, /company/windfarm/3/turbine/7/temperature). For more information, see Mapping industrial data streams to asset properties in the IoT SiteWise User Guide.

$sel:nextToken:GetAssetPropertyAggregates', getAssetPropertyAggregates_nextToken - The token to be used for the next set of paginated results.

$sel:propertyId:GetAssetPropertyAggregates', getAssetPropertyAggregates_propertyId - The ID of the asset property.

$sel:assetId:GetAssetPropertyAggregates', getAssetPropertyAggregates_assetId - The ID of the asset.

$sel:maxResults:GetAssetPropertyAggregates', getAssetPropertyAggregates_maxResults - The maximum number of results to return for each paginated request.

Default: 100

$sel:aggregateTypes:GetAssetPropertyAggregates', getAssetPropertyAggregates_aggregateTypes - The data aggregating function.

$sel:resolution:GetAssetPropertyAggregates', getAssetPropertyAggregates_resolution - The time interval over which to aggregate data.

$sel:startDate:GetAssetPropertyAggregates', getAssetPropertyAggregates_startDate - The exclusive start of the range from which to query historical data, expressed in seconds in Unix epoch time.

$sel:endDate:GetAssetPropertyAggregates', getAssetPropertyAggregates_endDate - The inclusive end of the range from which to query historical data, expressed in seconds in Unix epoch time.

Request Lenses

getAssetPropertyAggregates_qualities :: Lens' GetAssetPropertyAggregates (Maybe (NonEmpty Quality))

The quality by which to filter asset data.

getAssetPropertyAggregates_timeOrdering :: Lens' GetAssetPropertyAggregates (Maybe TimeOrdering)

The chronological sorting order of the requested information.

Default: ASCENDING

getAssetPropertyAggregates_propertyAlias :: Lens' GetAssetPropertyAggregates (Maybe Text)

The alias that identifies the property, such as an OPC-UA server data stream path (for example, /company/windfarm/3/turbine/7/temperature). For more information, see Mapping industrial data streams to asset properties in the IoT SiteWise User Guide.

getAssetPropertyAggregates_nextToken :: Lens' GetAssetPropertyAggregates (Maybe Text)

The token to be used for the next set of paginated results.

getAssetPropertyAggregates_propertyId :: Lens' GetAssetPropertyAggregates (Maybe Text)

The ID of the asset property.

getAssetPropertyAggregates_assetId :: Lens' GetAssetPropertyAggregates (Maybe Text)

The ID of the asset.

getAssetPropertyAggregates_maxResults :: Lens' GetAssetPropertyAggregates (Maybe Natural)

The maximum number of results to return for each paginated request.

Default: 100

getAssetPropertyAggregates_aggregateTypes :: Lens' GetAssetPropertyAggregates (NonEmpty AggregateType)

The data aggregating function.

getAssetPropertyAggregates_resolution :: Lens' GetAssetPropertyAggregates Text

The time interval over which to aggregate data.

getAssetPropertyAggregates_startDate :: Lens' GetAssetPropertyAggregates UTCTime

The exclusive start of the range from which to query historical data, expressed in seconds in Unix epoch time.

getAssetPropertyAggregates_endDate :: Lens' GetAssetPropertyAggregates UTCTime

The inclusive end of the range from which to query historical data, expressed in seconds in Unix epoch time.

Destructuring the Response

data GetAssetPropertyAggregatesResponse

See: newGetAssetPropertyAggregatesResponse smart constructor.

Constructors

GetAssetPropertyAggregatesResponse'

Fields

nextToken :: Maybe Text

The token for the next set of results, or null if there are no additional results.

httpStatus :: Int

The response's http status code.

aggregatedValues :: [AggregatedValue]

The requested aggregated values.

newGetAssetPropertyAggregatesResponse

Arguments

:: Int
$sel:httpStatus:GetAssetPropertyAggregatesResponse'

-> GetAssetPropertyAggregatesResponse

Create a value of GetAssetPropertyAggregatesResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:nextToken:GetAssetPropertyAggregates', getAssetPropertyAggregatesResponse_nextToken - The token for the next set of results, or null if there are no additional results.

$sel:httpStatus:GetAssetPropertyAggregatesResponse', getAssetPropertyAggregatesResponse_httpStatus - The response's http status code.

$sel:aggregatedValues:GetAssetPropertyAggregatesResponse', getAssetPropertyAggregatesResponse_aggregatedValues - The requested aggregated values.

Response Lenses

getAssetPropertyAggregatesResponse_nextToken :: Lens' GetAssetPropertyAggregatesResponse (Maybe Text)

The token for the next set of results, or null if there are no additional results.

getAssetPropertyAggregatesResponse_httpStatus :: Lens' GetAssetPropertyAggregatesResponse Int

The response's http status code.

getAssetPropertyAggregatesResponse_aggregatedValues :: Lens' GetAssetPropertyAggregatesResponse [AggregatedValue]

The requested aggregated values.