Amazonka.IoTSiteWise.GetInterpolatedAssetPropertyValues

Description

Get interpolated values for an asset property for a specified time interval, during a period of time. If your time series is missing data points during the specified time interval, you can use interpolation to estimate the missing data.

For example, you can use this operation to return the interpolated temperature values for a wind turbine every 24 hours over a duration of 7 days.

To identify an asset property, you must specify one of the following:

The assetId and propertyId of an asset property.

A propertyAlias, which is a data stream alias (for example, /company/windfarm/3/turbine/7/temperature). To define an asset property's alias, see UpdateAssetProperty.

This operation returns paginated results.

Creating a Request

data GetInterpolatedAssetPropertyValues

See: newGetInterpolatedAssetPropertyValues smart constructor.

Constructors

GetInterpolatedAssetPropertyValues'

Fields

startTimeOffsetInNanos :: Maybe Natural

The nanosecond offset converted from startTimeInSeconds.

endTimeOffsetInNanos :: Maybe Natural

The nanosecond offset converted from endTimeInSeconds.

propertyAlias :: Maybe Text

The alias that identifies the property, such as an OPC-UA server data stream path (for example, /company/windfarm/3/turbine/7/temperature). For more information, see Mapping industrial data streams to asset properties in the IoT SiteWise User Guide.

nextToken :: Maybe Text

The token to be used for the next set of paginated results.

intervalWindowInSeconds :: Maybe Natural

The query interval for the window in seconds. IoT SiteWise computes each interpolated value by using data points from the timestamp of each interval minus the window to the timestamp of each interval plus the window. If not specified, the window is between the start time minus the interval and the end time plus the interval.

If you specify a value for the intervalWindowInSeconds parameter, the type parameter must be LINEAR_INTERPOLATION.

If no data point is found during the specified query window, IoT SiteWise won't return an interpolated value for the interval. This indicates that there's a gap in the ingested data points.

For example, you can get the interpolated temperature values for a wind turbine every 24 hours over a duration of 7 days. If the interpolation starts on July 1, 2021, at 9 AM with a window of 2 hours, IoT SiteWise uses the data points from 7 AM (9 AM - 2 hours) to 11 AM (9 AM + 2 hours) on July 2, 2021 to compute the first interpolated value, uses the data points from 7 AM (9 AM - 2 hours) to 11 AM (9 AM + 2 hours) on July 3, 2021 to compute the second interpolated value, and so on.

propertyId :: Maybe Text

The ID of the asset property.

assetId :: Maybe Text

The ID of the asset.

maxResults :: Maybe Natural

The maximum number of results to return for each paginated request. If not specified, the default value is 10.

startTimeInSeconds :: Natural

The exclusive start of the range from which to interpolate data, expressed in seconds in Unix epoch time.

endTimeInSeconds :: Natural

The inclusive end of the range from which to interpolate data, expressed in seconds in Unix epoch time.

quality :: Quality

The quality of the asset property value. You can use this parameter as a filter to choose only the asset property values that have a specific quality.

intervalInSeconds :: Natural

The time interval in seconds over which to interpolate data. Each interval starts when the previous one ends.

type' :: Text

The interpolation type.

Valid values: LINEAR_INTERPOLATION | LOCF_INTERPOLATION

LINEAR_INTERPOLATION – Estimates missing data using linear interpolation.

For example, you can use this operation to return the interpolated temperature values for a wind turbine every 24 hours over a duration of 7 days. If the interpolation starts on July 1, 2021, at 9 AM, IoT SiteWise returns the first interpolated value on July 2, 2021, at 9 AM, the second interpolated value on July 3, 2021, at 9 AM, and so on.

LOCF_INTERPOLATION – Estimates missing data using last observation carried forward interpolation

If no data point is found for an interval, IoT SiteWise returns the last observed data point for the previous interval and carries forward this interpolated value until a new data point is found.

For example, you can get the state of an on-off valve every 24 hours over a duration of 7 days. If the interpolation starts on July 1, 2021, at 9 AM, IoT SiteWise returns the last observed data point between July 1, 2021, at 9 AM and July 2, 2021, at 9 AM as the first interpolated value. If no data point is found after 9 AM on July 2, 2021, IoT SiteWise uses the same interpolated value for the rest of the days.

newGetInterpolatedAssetPropertyValues

Arguments

:: Natural
$sel:startTimeInSeconds:GetInterpolatedAssetPropertyValues'

-> Natural
$sel:endTimeInSeconds:GetInterpolatedAssetPropertyValues'

-> Quality
$sel:quality:GetInterpolatedAssetPropertyValues'

-> Natural
$sel:intervalInSeconds:GetInterpolatedAssetPropertyValues'

-> Text
$sel:type':GetInterpolatedAssetPropertyValues'

-> GetInterpolatedAssetPropertyValues

Create a value of GetInterpolatedAssetPropertyValues with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:startTimeOffsetInNanos:GetInterpolatedAssetPropertyValues', getInterpolatedAssetPropertyValues_startTimeOffsetInNanos - The nanosecond offset converted from startTimeInSeconds.

$sel:endTimeOffsetInNanos:GetInterpolatedAssetPropertyValues', getInterpolatedAssetPropertyValues_endTimeOffsetInNanos - The nanosecond offset converted from endTimeInSeconds.

$sel:propertyAlias:GetInterpolatedAssetPropertyValues', getInterpolatedAssetPropertyValues_propertyAlias - The alias that identifies the property, such as an OPC-UA server data stream path (for example, /company/windfarm/3/turbine/7/temperature). For more information, see Mapping industrial data streams to asset properties in the IoT SiteWise User Guide.

$sel:nextToken:GetInterpolatedAssetPropertyValues', getInterpolatedAssetPropertyValues_nextToken - The token to be used for the next set of paginated results.

$sel:intervalWindowInSeconds:GetInterpolatedAssetPropertyValues', getInterpolatedAssetPropertyValues_intervalWindowInSeconds - The query interval for the window in seconds. IoT SiteWise computes each interpolated value by using data points from the timestamp of each interval minus the window to the timestamp of each interval plus the window. If not specified, the window is between the start time minus the interval and the end time plus the interval.

If you specify a value for the intervalWindowInSeconds parameter, the type parameter must be LINEAR_INTERPOLATION.

If no data point is found during the specified query window, IoT SiteWise won't return an interpolated value for the interval. This indicates that there's a gap in the ingested data points.

For example, you can get the interpolated temperature values for a wind turbine every 24 hours over a duration of 7 days. If the interpolation starts on July 1, 2021, at 9 AM with a window of 2 hours, IoT SiteWise uses the data points from 7 AM (9 AM - 2 hours) to 11 AM (9 AM + 2 hours) on July 2, 2021 to compute the first interpolated value, uses the data points from 7 AM (9 AM - 2 hours) to 11 AM (9 AM + 2 hours) on July 3, 2021 to compute the second interpolated value, and so on.

$sel:propertyId:GetInterpolatedAssetPropertyValues', getInterpolatedAssetPropertyValues_propertyId - The ID of the asset property.

$sel:assetId:GetInterpolatedAssetPropertyValues', getInterpolatedAssetPropertyValues_assetId - The ID of the asset.

$sel:maxResults:GetInterpolatedAssetPropertyValues', getInterpolatedAssetPropertyValues_maxResults - The maximum number of results to return for each paginated request. If not specified, the default value is 10.

$sel:startTimeInSeconds:GetInterpolatedAssetPropertyValues', getInterpolatedAssetPropertyValues_startTimeInSeconds - The exclusive start of the range from which to interpolate data, expressed in seconds in Unix epoch time.

$sel:endTimeInSeconds:GetInterpolatedAssetPropertyValues', getInterpolatedAssetPropertyValues_endTimeInSeconds - The inclusive end of the range from which to interpolate data, expressed in seconds in Unix epoch time.

$sel:quality:GetInterpolatedAssetPropertyValues', getInterpolatedAssetPropertyValues_quality - The quality of the asset property value. You can use this parameter as a filter to choose only the asset property values that have a specific quality.

$sel:intervalInSeconds:GetInterpolatedAssetPropertyValues', getInterpolatedAssetPropertyValues_intervalInSeconds - The time interval in seconds over which to interpolate data. Each interval starts when the previous one ends.

$sel:type':GetInterpolatedAssetPropertyValues', getInterpolatedAssetPropertyValues_type - The interpolation type.

Valid values: LINEAR_INTERPOLATION | LOCF_INTERPOLATION

LINEAR_INTERPOLATION – Estimates missing data using linear interpolation.

For example, you can use this operation to return the interpolated temperature values for a wind turbine every 24 hours over a duration of 7 days. If the interpolation starts on July 1, 2021, at 9 AM, IoT SiteWise returns the first interpolated value on July 2, 2021, at 9 AM, the second interpolated value on July 3, 2021, at 9 AM, and so on.

LOCF_INTERPOLATION – Estimates missing data using last observation carried forward interpolation

If no data point is found for an interval, IoT SiteWise returns the last observed data point for the previous interval and carries forward this interpolated value until a new data point is found.

For example, you can get the state of an on-off valve every 24 hours over a duration of 7 days. If the interpolation starts on July 1, 2021, at 9 AM, IoT SiteWise returns the last observed data point between July 1, 2021, at 9 AM and July 2, 2021, at 9 AM as the first interpolated value. If no data point is found after 9 AM on July 2, 2021, IoT SiteWise uses the same interpolated value for the rest of the days.

Request Lenses

getInterpolatedAssetPropertyValues_startTimeOffsetInNanos :: Lens' GetInterpolatedAssetPropertyValues (Maybe Natural)

The nanosecond offset converted from startTimeInSeconds.

getInterpolatedAssetPropertyValues_endTimeOffsetInNanos :: Lens' GetInterpolatedAssetPropertyValues (Maybe Natural)

The nanosecond offset converted from endTimeInSeconds.

getInterpolatedAssetPropertyValues_propertyAlias :: Lens' GetInterpolatedAssetPropertyValues (Maybe Text)

The alias that identifies the property, such as an OPC-UA server data stream path (for example, /company/windfarm/3/turbine/7/temperature). For more information, see Mapping industrial data streams to asset properties in the IoT SiteWise User Guide.

getInterpolatedAssetPropertyValues_nextToken :: Lens' GetInterpolatedAssetPropertyValues (Maybe Text)

The token to be used for the next set of paginated results.

getInterpolatedAssetPropertyValues_intervalWindowInSeconds :: Lens' GetInterpolatedAssetPropertyValues (Maybe Natural)

The query interval for the window in seconds. IoT SiteWise computes each interpolated value by using data points from the timestamp of each interval minus the window to the timestamp of each interval plus the window. If not specified, the window is between the start time minus the interval and the end time plus the interval.

If you specify a value for the intervalWindowInSeconds parameter, the type parameter must be LINEAR_INTERPOLATION.

If no data point is found during the specified query window, IoT SiteWise won't return an interpolated value for the interval. This indicates that there's a gap in the ingested data points.

For example, you can get the interpolated temperature values for a wind turbine every 24 hours over a duration of 7 days. If the interpolation starts on July 1, 2021, at 9 AM with a window of 2 hours, IoT SiteWise uses the data points from 7 AM (9 AM - 2 hours) to 11 AM (9 AM + 2 hours) on July 2, 2021 to compute the first interpolated value, uses the data points from 7 AM (9 AM - 2 hours) to 11 AM (9 AM + 2 hours) on July 3, 2021 to compute the second interpolated value, and so on.

getInterpolatedAssetPropertyValues_propertyId :: Lens' GetInterpolatedAssetPropertyValues (Maybe Text)

The ID of the asset property.

getInterpolatedAssetPropertyValues_assetId :: Lens' GetInterpolatedAssetPropertyValues (Maybe Text)

The ID of the asset.

getInterpolatedAssetPropertyValues_maxResults :: Lens' GetInterpolatedAssetPropertyValues (Maybe Natural)

The maximum number of results to return for each paginated request. If not specified, the default value is 10.

getInterpolatedAssetPropertyValues_startTimeInSeconds :: Lens' GetInterpolatedAssetPropertyValues Natural

The exclusive start of the range from which to interpolate data, expressed in seconds in Unix epoch time.

getInterpolatedAssetPropertyValues_endTimeInSeconds :: Lens' GetInterpolatedAssetPropertyValues Natural

The inclusive end of the range from which to interpolate data, expressed in seconds in Unix epoch time.

getInterpolatedAssetPropertyValues_quality :: Lens' GetInterpolatedAssetPropertyValues Quality

The quality of the asset property value. You can use this parameter as a filter to choose only the asset property values that have a specific quality.

getInterpolatedAssetPropertyValues_intervalInSeconds :: Lens' GetInterpolatedAssetPropertyValues Natural

The time interval in seconds over which to interpolate data. Each interval starts when the previous one ends.

getInterpolatedAssetPropertyValues_type :: Lens' GetInterpolatedAssetPropertyValues Text

The interpolation type.

Valid values: LINEAR_INTERPOLATION | LOCF_INTERPOLATION

LINEAR_INTERPOLATION – Estimates missing data using linear interpolation.

For example, you can use this operation to return the interpolated temperature values for a wind turbine every 24 hours over a duration of 7 days. If the interpolation starts on July 1, 2021, at 9 AM, IoT SiteWise returns the first interpolated value on July 2, 2021, at 9 AM, the second interpolated value on July 3, 2021, at 9 AM, and so on.

LOCF_INTERPOLATION – Estimates missing data using last observation carried forward interpolation

If no data point is found for an interval, IoT SiteWise returns the last observed data point for the previous interval and carries forward this interpolated value until a new data point is found.

For example, you can get the state of an on-off valve every 24 hours over a duration of 7 days. If the interpolation starts on July 1, 2021, at 9 AM, IoT SiteWise returns the last observed data point between July 1, 2021, at 9 AM and July 2, 2021, at 9 AM as the first interpolated value. If no data point is found after 9 AM on July 2, 2021, IoT SiteWise uses the same interpolated value for the rest of the days.

Destructuring the Response

data GetInterpolatedAssetPropertyValuesResponse

See: newGetInterpolatedAssetPropertyValuesResponse smart constructor.

Constructors

GetInterpolatedAssetPropertyValuesResponse'

Fields

nextToken :: Maybe Text

The token for the next set of results, or null if there are no additional results.

httpStatus :: Int

The response's http status code.

interpolatedAssetPropertyValues :: [InterpolatedAssetPropertyValue]

The requested interpolated values.

newGetInterpolatedAssetPropertyValuesResponse

Arguments

:: Int
$sel:httpStatus:GetInterpolatedAssetPropertyValuesResponse'

-> GetInterpolatedAssetPropertyValuesResponse

Create a value of GetInterpolatedAssetPropertyValuesResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:nextToken:GetInterpolatedAssetPropertyValues', getInterpolatedAssetPropertyValuesResponse_nextToken - The token for the next set of results, or null if there are no additional results.

$sel:httpStatus:GetInterpolatedAssetPropertyValuesResponse', getInterpolatedAssetPropertyValuesResponse_httpStatus - The response's http status code.

$sel:interpolatedAssetPropertyValues:GetInterpolatedAssetPropertyValuesResponse', getInterpolatedAssetPropertyValuesResponse_interpolatedAssetPropertyValues - The requested interpolated values.

Response Lenses

getInterpolatedAssetPropertyValuesResponse_nextToken :: Lens' GetInterpolatedAssetPropertyValuesResponse (Maybe Text)

The token for the next set of results, or null if there are no additional results.

getInterpolatedAssetPropertyValuesResponse_httpStatus :: Lens' GetInterpolatedAssetPropertyValuesResponse Int

The response's http status code.

getInterpolatedAssetPropertyValuesResponse_interpolatedAssetPropertyValues :: Lens' GetInterpolatedAssetPropertyValuesResponse [InterpolatedAssetPropertyValue]

The requested interpolated values.