Amazonka.IoTSiteWise.ListAccessPolicies

Description

Retrieves a paginated list of access policies for an identity (an Amazon Web Services SSO user, an Amazon Web Services SSO group, or an IAM user) or an IoT SiteWise Monitor resource (a portal or project).

This operation returns paginated results.

Creating a Request

data ListAccessPolicies

See: newListAccessPolicies smart constructor.

Constructors

ListAccessPolicies'

Fields

resourceId :: Maybe Text

The ID of the resource. This parameter is required if you specify resourceType.

resourceType :: Maybe ResourceType

The type of resource (portal or project). This parameter is required if you specify resourceId.

identityType :: Maybe IdentityType

The type of identity (Amazon Web Services SSO user, Amazon Web Services SSO group, or IAM user). This parameter is required if you specify identityId.

nextToken :: Maybe Text

The token to be used for the next set of paginated results.

iamArn :: Maybe Text

The ARN of the IAM user. For more information, see IAM ARNs in the IAM User Guide. This parameter is required if you specify IAM for identityType.

identityId :: Maybe Text

The ID of the identity. This parameter is required if you specify USER or GROUP for identityType.

maxResults :: Maybe Natural

The maximum number of results to return for each paginated request.

Default: 50

newListAccessPolicies :: ListAccessPolicies

Create a value of ListAccessPolicies with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:resourceId:ListAccessPolicies', listAccessPolicies_resourceId - The ID of the resource. This parameter is required if you specify resourceType.

$sel:resourceType:ListAccessPolicies', listAccessPolicies_resourceType - The type of resource (portal or project). This parameter is required if you specify resourceId.

$sel:identityType:ListAccessPolicies', listAccessPolicies_identityType - The type of identity (Amazon Web Services SSO user, Amazon Web Services SSO group, or IAM user). This parameter is required if you specify identityId.

$sel:nextToken:ListAccessPolicies', listAccessPolicies_nextToken - The token to be used for the next set of paginated results.

$sel:iamArn:ListAccessPolicies', listAccessPolicies_iamArn - The ARN of the IAM user. For more information, see IAM ARNs in the IAM User Guide. This parameter is required if you specify IAM for identityType.

$sel:identityId:ListAccessPolicies', listAccessPolicies_identityId - The ID of the identity. This parameter is required if you specify USER or GROUP for identityType.

$sel:maxResults:ListAccessPolicies', listAccessPolicies_maxResults - The maximum number of results to return for each paginated request.

Default: 50

Request Lenses

listAccessPolicies_resourceId :: Lens' ListAccessPolicies (Maybe Text)

The ID of the resource. This parameter is required if you specify resourceType.

listAccessPolicies_resourceType :: Lens' ListAccessPolicies (Maybe ResourceType)

The type of resource (portal or project). This parameter is required if you specify resourceId.

listAccessPolicies_identityType :: Lens' ListAccessPolicies (Maybe IdentityType)

The type of identity (Amazon Web Services SSO user, Amazon Web Services SSO group, or IAM user). This parameter is required if you specify identityId.

listAccessPolicies_nextToken :: Lens' ListAccessPolicies (Maybe Text)

The token to be used for the next set of paginated results.

listAccessPolicies_iamArn :: Lens' ListAccessPolicies (Maybe Text)

The ARN of the IAM user. For more information, see IAM ARNs in the IAM User Guide. This parameter is required if you specify IAM for identityType.

listAccessPolicies_identityId :: Lens' ListAccessPolicies (Maybe Text)

The ID of the identity. This parameter is required if you specify USER or GROUP for identityType.

listAccessPolicies_maxResults :: Lens' ListAccessPolicies (Maybe Natural)

The maximum number of results to return for each paginated request.

Default: 50

Destructuring the Response

data ListAccessPoliciesResponse

See: newListAccessPoliciesResponse smart constructor.

Constructors

ListAccessPoliciesResponse'

Fields

nextToken :: Maybe Text

The token for the next set of results, or null if there are no additional results.

httpStatus :: Int

The response's http status code.

accessPolicySummaries :: [AccessPolicySummary]

A list that summarizes each access policy.

newListAccessPoliciesResponse

Arguments

:: Int
$sel:httpStatus:ListAccessPoliciesResponse'

-> ListAccessPoliciesResponse

Create a value of ListAccessPoliciesResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:nextToken:ListAccessPolicies', listAccessPoliciesResponse_nextToken - The token for the next set of results, or null if there are no additional results.

$sel:httpStatus:ListAccessPoliciesResponse', listAccessPoliciesResponse_httpStatus - The response's http status code.

$sel:accessPolicySummaries:ListAccessPoliciesResponse', listAccessPoliciesResponse_accessPolicySummaries - A list that summarizes each access policy.

Response Lenses

listAccessPoliciesResponse_nextToken :: Lens' ListAccessPoliciesResponse (Maybe Text)

The token for the next set of results, or null if there are no additional results.

listAccessPoliciesResponse_httpStatus :: Lens' ListAccessPoliciesResponse Int

The response's http status code.

listAccessPoliciesResponse_accessPolicySummaries :: Lens' ListAccessPoliciesResponse [AccessPolicySummary]

A list that summarizes each access policy.