Amazonka.IoTSiteWise.ListAssociatedAssets

Description

Retrieves a paginated list of associated assets.

You can use this operation to do the following:

List child assets associated to a parent asset by a hierarchy that you specify.

List an asset's parent asset.

This operation returns paginated results.

Creating a Request

data ListAssociatedAssets

See: newListAssociatedAssets smart constructor.

Constructors

ListAssociatedAssets'

Fields

hierarchyId :: Maybe Text

The ID of the hierarchy by which child assets are associated to the asset. To find a hierarchy ID, use the DescribeAsset or DescribeAssetModel operations. This parameter is required if you choose CHILD for traversalDirection.

For more information, see Asset hierarchies in the IoT SiteWise User Guide.

traversalDirection :: Maybe TraversalDirection

The direction to list associated assets. Choose one of the following options:

CHILD – The list includes all child assets associated to the asset. The hierarchyId parameter is required if you choose CHILD.

PARENT – The list includes the asset's parent asset.

Default: CHILD

nextToken :: Maybe Text

The token to be used for the next set of paginated results.

maxResults :: Maybe Natural

The maximum number of results to return for each paginated request.

Default: 50

assetId :: Text

The ID of the asset to query.

newListAssociatedAssets

Arguments

:: Text
$sel:assetId:ListAssociatedAssets'

-> ListAssociatedAssets

Create a value of ListAssociatedAssets with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:hierarchyId:ListAssociatedAssets', listAssociatedAssets_hierarchyId - The ID of the hierarchy by which child assets are associated to the asset. To find a hierarchy ID, use the DescribeAsset or DescribeAssetModel operations. This parameter is required if you choose CHILD for traversalDirection.

For more information, see Asset hierarchies in the IoT SiteWise User Guide.

$sel:traversalDirection:ListAssociatedAssets', listAssociatedAssets_traversalDirection - The direction to list associated assets. Choose one of the following options:

CHILD – The list includes all child assets associated to the asset. The hierarchyId parameter is required if you choose CHILD.

PARENT – The list includes the asset's parent asset.

Default: CHILD

$sel:nextToken:ListAssociatedAssets', listAssociatedAssets_nextToken - The token to be used for the next set of paginated results.

$sel:maxResults:ListAssociatedAssets', listAssociatedAssets_maxResults - The maximum number of results to return for each paginated request.

Default: 50

$sel:assetId:ListAssociatedAssets', listAssociatedAssets_assetId - The ID of the asset to query.

Request Lenses

listAssociatedAssets_hierarchyId :: Lens' ListAssociatedAssets (Maybe Text)

The ID of the hierarchy by which child assets are associated to the asset. To find a hierarchy ID, use the DescribeAsset or DescribeAssetModel operations. This parameter is required if you choose CHILD for traversalDirection.

For more information, see Asset hierarchies in the IoT SiteWise User Guide.

listAssociatedAssets_traversalDirection :: Lens' ListAssociatedAssets (Maybe TraversalDirection)

The direction to list associated assets. Choose one of the following options:

CHILD – The list includes all child assets associated to the asset. The hierarchyId parameter is required if you choose CHILD.

PARENT – The list includes the asset's parent asset.

Default: CHILD

listAssociatedAssets_nextToken :: Lens' ListAssociatedAssets (Maybe Text)

The token to be used for the next set of paginated results.

listAssociatedAssets_maxResults :: Lens' ListAssociatedAssets (Maybe Natural)

The maximum number of results to return for each paginated request.

Default: 50

listAssociatedAssets_assetId :: Lens' ListAssociatedAssets Text

The ID of the asset to query.

Destructuring the Response

data ListAssociatedAssetsResponse

See: newListAssociatedAssetsResponse smart constructor.

Constructors

ListAssociatedAssetsResponse'

Fields

nextToken :: Maybe Text

The token for the next set of results, or null if there are no additional results.

httpStatus :: Int

The response's http status code.

assetSummaries :: [AssociatedAssetsSummary]

A list that summarizes the associated assets.

newListAssociatedAssetsResponse

Arguments

:: Int
$sel:httpStatus:ListAssociatedAssetsResponse'

-> ListAssociatedAssetsResponse

Create a value of ListAssociatedAssetsResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:nextToken:ListAssociatedAssets', listAssociatedAssetsResponse_nextToken - The token for the next set of results, or null if there are no additional results.

$sel:httpStatus:ListAssociatedAssetsResponse', listAssociatedAssetsResponse_httpStatus - The response's http status code.

$sel:assetSummaries:ListAssociatedAssetsResponse', listAssociatedAssetsResponse_assetSummaries - A list that summarizes the associated assets.

Response Lenses

listAssociatedAssetsResponse_nextToken :: Lens' ListAssociatedAssetsResponse (Maybe Text)

The token for the next set of results, or null if there are no additional results.

listAssociatedAssetsResponse_httpStatus :: Lens' ListAssociatedAssetsResponse Int

The response's http status code.

listAssociatedAssetsResponse_assetSummaries :: Lens' ListAssociatedAssetsResponse [AssociatedAssetsSummary]

A list that summarizes the associated assets.