Amazonka.IoTSiteWise.ListPortals

Description

Retrieves a paginated list of IoT SiteWise Monitor portals.

This operation returns paginated results.

Creating a Request

data ListPortals

See: newListPortals smart constructor.

Constructors

ListPortals'

Fields

nextToken :: Maybe Text

The token to be used for the next set of paginated results.

maxResults :: Maybe Natural

The maximum number of results to return for each paginated request.

Default: 50

newListPortals :: ListPortals

Create a value of ListPortals with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:nextToken:ListPortals', listPortals_nextToken - The token to be used for the next set of paginated results.

$sel:maxResults:ListPortals', listPortals_maxResults - The maximum number of results to return for each paginated request.

Default: 50

Request Lenses

listPortals_nextToken :: Lens' ListPortals (Maybe Text)

The token to be used for the next set of paginated results.

listPortals_maxResults :: Lens' ListPortals (Maybe Natural)

The maximum number of results to return for each paginated request.

Default: 50

Destructuring the Response

data ListPortalsResponse

See: newListPortalsResponse smart constructor.

Constructors

ListPortalsResponse'

Fields

portalSummaries :: Maybe [PortalSummary]

A list that summarizes each portal.

nextToken :: Maybe Text

The token for the next set of results, or null if there are no additional results.

httpStatus :: Int

The response's http status code.

newListPortalsResponse

Arguments

:: Int
$sel:httpStatus:ListPortalsResponse'

-> ListPortalsResponse

Create a value of ListPortalsResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:portalSummaries:ListPortalsResponse', listPortalsResponse_portalSummaries - A list that summarizes each portal.

$sel:nextToken:ListPortals', listPortalsResponse_nextToken - The token for the next set of results, or null if there are no additional results.

$sel:httpStatus:ListPortalsResponse', listPortalsResponse_httpStatus - The response's http status code.

Response Lenses

listPortalsResponse_portalSummaries :: Lens' ListPortalsResponse (Maybe [PortalSummary])

A list that summarizes each portal.

listPortalsResponse_nextToken :: Lens' ListPortalsResponse (Maybe Text)

The token for the next set of results, or null if there are no additional results.

listPortalsResponse_httpStatus :: Lens' ListPortalsResponse Int

The response's http status code.