Amazonka.IoTSiteWise.Types.AccessPolicySummary

Description

Documentation

data AccessPolicySummary

Contains an access policy that defines an identity's access to an IoT SiteWise Monitor resource.

See: newAccessPolicySummary smart constructor.

Constructors

AccessPolicySummary'

Fields

lastUpdateDate :: Maybe POSIX

The date the access policy was last updated, in Unix epoch time.

creationDate :: Maybe POSIX

The date the access policy was created, in Unix epoch time.

id :: Text

The ID of the access policy.

identity :: Identity

The identity (an Amazon Web Services SSO user, an Amazon Web Services SSO group, or an IAM user).

resource :: Resource

The IoT SiteWise Monitor resource (a portal or project).

permission :: Permission

The permissions for the access policy. Note that a project ADMINISTRATOR is also known as a project owner.

newAccessPolicySummary

Arguments

:: Text
$sel:id:AccessPolicySummary'

-> Identity
$sel:identity:AccessPolicySummary'

-> Resource
$sel:resource:AccessPolicySummary'

-> Permission
$sel:permission:AccessPolicySummary'

-> AccessPolicySummary

Create a value of AccessPolicySummary with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:lastUpdateDate:AccessPolicySummary', accessPolicySummary_lastUpdateDate - The date the access policy was last updated, in Unix epoch time.

$sel:creationDate:AccessPolicySummary', accessPolicySummary_creationDate - The date the access policy was created, in Unix epoch time.

$sel:id:AccessPolicySummary', accessPolicySummary_id - The ID of the access policy.

$sel:identity:AccessPolicySummary', accessPolicySummary_identity - The identity (an Amazon Web Services SSO user, an Amazon Web Services SSO group, or an IAM user).

$sel:resource:AccessPolicySummary', accessPolicySummary_resource - The IoT SiteWise Monitor resource (a portal or project).

$sel:permission:AccessPolicySummary', accessPolicySummary_permission - The permissions for the access policy. Note that a project ADMINISTRATOR is also known as a project owner.

accessPolicySummary_lastUpdateDate :: Lens' AccessPolicySummary (Maybe UTCTime)

The date the access policy was last updated, in Unix epoch time.

accessPolicySummary_creationDate :: Lens' AccessPolicySummary (Maybe UTCTime)

The date the access policy was created, in Unix epoch time.

accessPolicySummary_id :: Lens' AccessPolicySummary Text

The ID of the access policy.

accessPolicySummary_identity :: Lens' AccessPolicySummary Identity

The identity (an Amazon Web Services SSO user, an Amazon Web Services SSO group, or an IAM user).

accessPolicySummary_resource :: Lens' AccessPolicySummary Resource

The IoT SiteWise Monitor resource (a portal or project).

accessPolicySummary_permission :: Lens' AccessPolicySummary Permission

The permissions for the access policy. Note that a project ADMINISTRATOR is also known as a project owner.