Amazonka.IoTSiteWise.Types.AggregatedValue

Description

Documentation

data AggregatedValue

Contains aggregated asset property values (for example, average, minimum, and maximum).

See: newAggregatedValue smart constructor.

Constructors

AggregatedValue'

Fields

quality :: Maybe Quality

The quality of the aggregated data.

timestamp :: POSIX

The date the aggregating computations occurred, in Unix epoch time.

value :: Aggregates

The value of the aggregates.

newAggregatedValue

Arguments

:: UTCTime
$sel:timestamp:AggregatedValue'

-> Aggregates
$sel:value:AggregatedValue'

-> AggregatedValue

Create a value of AggregatedValue with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:quality:AggregatedValue', aggregatedValue_quality - The quality of the aggregated data.

$sel:timestamp:AggregatedValue', aggregatedValue_timestamp - The date the aggregating computations occurred, in Unix epoch time.

$sel:value:AggregatedValue', aggregatedValue_value - The value of the aggregates.

aggregatedValue_quality :: Lens' AggregatedValue (Maybe Quality)

The quality of the aggregated data.

aggregatedValue_timestamp :: Lens' AggregatedValue UTCTime

The date the aggregating computations occurred, in Unix epoch time.

aggregatedValue_value :: Lens' AggregatedValue Aggregates

The value of the aggregates.