Amazonka.IoTSiteWise.Types.Alarms

Description

Documentation

data Alarms

Contains the configuration information of an alarm created in an IoT SiteWise Monitor portal. You can use the alarm to monitor an asset property and get notified when the asset property value is outside a specified range. For more information, see Monitoring with alarms in the IoT SiteWise Application Guide.

See: newAlarms smart constructor.

Constructors

Alarms'

Fields

notificationLambdaArn :: Maybe Text

The ARN of the Lambda function that manages alarm notifications. For more information, see Managing alarm notifications in the IoT Events Developer Guide.

alarmRoleArn :: Text

The ARN of the IAM role that allows the alarm to perform actions and access Amazon Web Services resources and services, such as IoT Events.

newAlarms

Arguments

:: Text
$sel:alarmRoleArn:Alarms'

-> Alarms

Create a value of Alarms with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:notificationLambdaArn:Alarms', alarms_notificationLambdaArn - The ARN of the Lambda function that manages alarm notifications. For more information, see Managing alarm notifications in the IoT Events Developer Guide.

$sel:alarmRoleArn:Alarms', alarms_alarmRoleArn - The ARN of the IAM role that allows the alarm to perform actions and access Amazon Web Services resources and services, such as IoT Events.

alarms_notificationLambdaArn :: Lens' Alarms (Maybe Text)

The ARN of the Lambda function that manages alarm notifications. For more information, see Managing alarm notifications in the IoT Events Developer Guide.

alarms_alarmRoleArn :: Lens' Alarms Text

The ARN of the IAM role that allows the alarm to perform actions and access Amazon Web Services resources and services, such as IoT Events.