Amazonka.IoTSiteWise.Types.AssociatedAssetsSummary

Description

Documentation

data AssociatedAssetsSummary

Contains a summary of an associated asset.

See: newAssociatedAssetsSummary smart constructor.

Constructors

AssociatedAssetsSummary'

Fields

id :: Text

The ID of the asset.

arn :: Text

The ARN of the asset, which has the following format.

arn:${Partition}:iotsitewise:${Region}:${Account}:asset/${AssetId}

name :: Text

The name of the asset.

assetModelId :: Text

The ID of the asset model used to create the asset.

creationDate :: POSIX

The date the asset was created, in Unix epoch time.

lastUpdateDate :: POSIX

The date the asset was last updated, in Unix epoch time.

status :: AssetStatus

The current status of the asset.

hierarchies :: [AssetHierarchy]

A list of asset hierarchies that each contain a hierarchyId. A hierarchy specifies allowed parent/child asset relationships.

newAssociatedAssetsSummary

Arguments

:: Text
$sel:id:AssociatedAssetsSummary'

-> Text
$sel:arn:AssociatedAssetsSummary'

-> Text
$sel:name:AssociatedAssetsSummary'

-> Text
$sel:assetModelId:AssociatedAssetsSummary'

-> UTCTime
$sel:creationDate:AssociatedAssetsSummary'

-> UTCTime
$sel:lastUpdateDate:AssociatedAssetsSummary'

-> AssetStatus
$sel:status:AssociatedAssetsSummary'

-> AssociatedAssetsSummary

Create a value of AssociatedAssetsSummary with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:id:AssociatedAssetsSummary', associatedAssetsSummary_id - The ID of the asset.

$sel:arn:AssociatedAssetsSummary', associatedAssetsSummary_arn - The ARN of the asset, which has the following format.

arn:${Partition}:iotsitewise:${Region}:${Account}:asset/${AssetId}

$sel:name:AssociatedAssetsSummary', associatedAssetsSummary_name - The name of the asset.

$sel:assetModelId:AssociatedAssetsSummary', associatedAssetsSummary_assetModelId - The ID of the asset model used to create the asset.

$sel:creationDate:AssociatedAssetsSummary', associatedAssetsSummary_creationDate - The date the asset was created, in Unix epoch time.

$sel:lastUpdateDate:AssociatedAssetsSummary', associatedAssetsSummary_lastUpdateDate - The date the asset was last updated, in Unix epoch time.

$sel:status:AssociatedAssetsSummary', associatedAssetsSummary_status - The current status of the asset.

$sel:hierarchies:AssociatedAssetsSummary', associatedAssetsSummary_hierarchies - A list of asset hierarchies that each contain a hierarchyId. A hierarchy specifies allowed parent/child asset relationships.

associatedAssetsSummary_id :: Lens' AssociatedAssetsSummary Text

The ID of the asset.

associatedAssetsSummary_arn :: Lens' AssociatedAssetsSummary Text

The ARN of the asset, which has the following format.

arn:${Partition}:iotsitewise:${Region}:${Account}:asset/${AssetId}

associatedAssetsSummary_name :: Lens' AssociatedAssetsSummary Text

The name of the asset.

associatedAssetsSummary_assetModelId :: Lens' AssociatedAssetsSummary Text

The ID of the asset model used to create the asset.

associatedAssetsSummary_creationDate :: Lens' AssociatedAssetsSummary UTCTime

The date the asset was created, in Unix epoch time.

associatedAssetsSummary_lastUpdateDate :: Lens' AssociatedAssetsSummary UTCTime

The date the asset was last updated, in Unix epoch time.

associatedAssetsSummary_status :: Lens' AssociatedAssetsSummary AssetStatus

The current status of the asset.

associatedAssetsSummary_hierarchies :: Lens' AssociatedAssetsSummary [AssetHierarchy]

A list of asset hierarchies that each contain a hierarchyId. A hierarchy specifies allowed parent/child asset relationships.