Amazonka.IoTSiteWise.Types.GatewayCapabilitySummary

Description

Documentation

data GatewayCapabilitySummary

Contains a summary of a gateway capability configuration.

See: newGatewayCapabilitySummary smart constructor.

Constructors

GatewayCapabilitySummary'

Fields

capabilityNamespace :: Text

The namespace of the capability configuration. For example, if you configure OPC-UA sources from the IoT SiteWise console, your OPC-UA capability configuration has the namespace iotsitewise:opcuacollector:version, where version is a number such as 1.

capabilitySyncStatus :: CapabilitySyncStatus

The synchronization status of the capability configuration. The sync status can be one of the following:

IN_SYNC – The gateway is running the capability configuration.

OUT_OF_SYNC – The gateway hasn't received the capability configuration.

SYNC_FAILED – The gateway rejected the capability configuration.

newGatewayCapabilitySummary

Arguments

:: Text
$sel:capabilityNamespace:GatewayCapabilitySummary'

-> CapabilitySyncStatus
$sel:capabilitySyncStatus:GatewayCapabilitySummary'

-> GatewayCapabilitySummary

Create a value of GatewayCapabilitySummary with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:capabilityNamespace:GatewayCapabilitySummary', gatewayCapabilitySummary_capabilityNamespace - The namespace of the capability configuration. For example, if you configure OPC-UA sources from the IoT SiteWise console, your OPC-UA capability configuration has the namespace iotsitewise:opcuacollector:version, where version is a number such as 1.

$sel:capabilitySyncStatus:GatewayCapabilitySummary', gatewayCapabilitySummary_capabilitySyncStatus - The synchronization status of the capability configuration. The sync status can be one of the following:

IN_SYNC – The gateway is running the capability configuration.

OUT_OF_SYNC – The gateway hasn't received the capability configuration.

SYNC_FAILED – The gateway rejected the capability configuration.

gatewayCapabilitySummary_capabilityNamespace :: Lens' GatewayCapabilitySummary Text

The namespace of the capability configuration. For example, if you configure OPC-UA sources from the IoT SiteWise console, your OPC-UA capability configuration has the namespace iotsitewise:opcuacollector:version, where version is a number such as 1.

gatewayCapabilitySummary_capabilitySyncStatus :: Lens' GatewayCapabilitySummary CapabilitySyncStatus

The synchronization status of the capability configuration. The sync status can be one of the following:

IN_SYNC – The gateway is running the capability configuration.

OUT_OF_SYNC – The gateway hasn't received the capability configuration.

SYNC_FAILED – The gateway rejected the capability configuration.