Amazonka.IoTSiteWise.Types.Image

Description

Documentation

data Image

Contains an image that is one of the following:

An image file. Choose this option to upload a new image.

The ID of an existing image. Choose this option to keep an existing image.

See: newImage smart constructor.

Constructors

Image'

Fields

id :: Maybe Text

The ID of an existing image. Specify this parameter to keep an existing image.

file :: Maybe ImageFile

newImage :: Image

Create a value of Image with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:id:Image', image_id - The ID of an existing image. Specify this parameter to keep an existing image.

$sel:file:Image', image_file - Undocumented member.

image_id :: Lens' Image (Maybe Text)

The ID of an existing image. Specify this parameter to keep an existing image.

image_file :: Lens' Image (Maybe ImageFile)

Undocumented member.