Amazonka.IoTSiteWise.UpdatePortal

Description

Updates an IoT SiteWise Monitor portal.

Creating a Request

data UpdatePortal

See: newUpdatePortal smart constructor.

Constructors

UpdatePortal'

Fields

clientToken :: Maybe Text

A unique case-sensitive identifier that you can provide to ensure the idempotency of the request. Don't reuse this client token if a new idempotent request is required.

portalDescription :: Maybe Text

A new description for the portal.

notificationSenderEmail :: Maybe Text

The email address that sends alarm notifications.

portalLogoImage :: Maybe Image

alarms :: Maybe Alarms

Contains the configuration information of an alarm created in an IoT SiteWise Monitor portal. You can use the alarm to monitor an asset property and get notified when the asset property value is outside a specified range. For more information, see Monitoring with alarms in the IoT SiteWise Application Guide.

portalId :: Text

The ID of the portal to update.

portalName :: Text

A new friendly name for the portal.

portalContactEmail :: Text

The Amazon Web Services administrator's contact email address.

roleArn :: Text

The ARN of a service role that allows the portal's users to access your IoT SiteWise resources on your behalf. For more information, see Using service roles for IoT SiteWise Monitor in the IoT SiteWise User Guide.

newUpdatePortal

Arguments

:: Text
$sel:portalId:UpdatePortal'

-> Text
$sel:portalName:UpdatePortal'

-> Text
$sel:portalContactEmail:UpdatePortal'

-> Text
$sel:roleArn:UpdatePortal'

-> UpdatePortal

Create a value of UpdatePortal with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:clientToken:UpdatePortal', updatePortal_clientToken - A unique case-sensitive identifier that you can provide to ensure the idempotency of the request. Don't reuse this client token if a new idempotent request is required.

$sel:portalDescription:UpdatePortal', updatePortal_portalDescription - A new description for the portal.

$sel:notificationSenderEmail:UpdatePortal', updatePortal_notificationSenderEmail - The email address that sends alarm notifications.

$sel:portalLogoImage:UpdatePortal', updatePortal_portalLogoImage - Undocumented member.

$sel:alarms:UpdatePortal', updatePortal_alarms - Contains the configuration information of an alarm created in an IoT SiteWise Monitor portal. You can use the alarm to monitor an asset property and get notified when the asset property value is outside a specified range. For more information, see Monitoring with alarms in the IoT SiteWise Application Guide.

$sel:portalId:UpdatePortal', updatePortal_portalId - The ID of the portal to update.

$sel:portalName:UpdatePortal', updatePortal_portalName - A new friendly name for the portal.

$sel:portalContactEmail:UpdatePortal', updatePortal_portalContactEmail - The Amazon Web Services administrator's contact email address.

$sel:roleArn:UpdatePortal', updatePortal_roleArn - The ARN of a service role that allows the portal's users to access your IoT SiteWise resources on your behalf. For more information, see Using service roles for IoT SiteWise Monitor in the IoT SiteWise User Guide.

Request Lenses

updatePortal_clientToken :: Lens' UpdatePortal (Maybe Text)

A unique case-sensitive identifier that you can provide to ensure the idempotency of the request. Don't reuse this client token if a new idempotent request is required.

updatePortal_portalDescription :: Lens' UpdatePortal (Maybe Text)

A new description for the portal.

updatePortal_notificationSenderEmail :: Lens' UpdatePortal (Maybe Text)

The email address that sends alarm notifications.

updatePortal_portalLogoImage :: Lens' UpdatePortal (Maybe Image)

Undocumented member.

updatePortal_alarms :: Lens' UpdatePortal (Maybe Alarms)

Contains the configuration information of an alarm created in an IoT SiteWise Monitor portal. You can use the alarm to monitor an asset property and get notified when the asset property value is outside a specified range. For more information, see Monitoring with alarms in the IoT SiteWise Application Guide.

updatePortal_portalId :: Lens' UpdatePortal Text

The ID of the portal to update.

updatePortal_portalName :: Lens' UpdatePortal Text

A new friendly name for the portal.

updatePortal_portalContactEmail :: Lens' UpdatePortal Text

The Amazon Web Services administrator's contact email address.

updatePortal_roleArn :: Lens' UpdatePortal Text

The ARN of a service role that allows the portal's users to access your IoT SiteWise resources on your behalf. For more information, see Using service roles for IoT SiteWise Monitor in the IoT SiteWise User Guide.

Destructuring the Response

data UpdatePortalResponse

See: newUpdatePortalResponse smart constructor.

Constructors

UpdatePortalResponse'

Fields

httpStatus :: Int

The response's http status code.

portalStatus :: PortalStatus

The status of the portal, which contains a state (UPDATING after successfully calling this operation) and any error message.

newUpdatePortalResponse

Arguments

:: Int
$sel:httpStatus:UpdatePortalResponse'

-> PortalStatus
$sel:portalStatus:UpdatePortalResponse'

-> UpdatePortalResponse

Create a value of UpdatePortalResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:httpStatus:UpdatePortalResponse', updatePortalResponse_httpStatus - The response's http status code.

$sel:portalStatus:UpdatePortalResponse', updatePortalResponse_portalStatus - The status of the portal, which contains a state (UPDATING after successfully calling this operation) and any error message.

Response Lenses

updatePortalResponse_httpStatus :: Lens' UpdatePortalResponse Int

The response's http status code.

updatePortalResponse_portalStatus :: Lens' UpdatePortalResponse PortalStatus

The status of the portal, which contains a state (UPDATING after successfully calling this operation) and any error message.