Amazonka.IoTThingsGraph.CreateSystemInstance

Description

Creates a system instance.

This action validates the system instance, prepares the deployment-related resources. For Greengrass deployments, it updates the Greengrass group that is specified by the greengrassGroupName parameter. It also adds a file to the S3 bucket specified by the s3BucketName parameter. You need to call DeploySystemInstance after running this action.

For Greengrass deployments, since this action modifies and adds resources to a Greengrass group and an S3 bucket on the caller's behalf, the calling identity must have write permissions to both the specified Greengrass group and S3 bucket. Otherwise, the call will fail with an authorization error.

For cloud deployments, this action requires a flowActionsRoleArn value. This is an IAM role that has permissions to access AWS services, such as AWS Lambda and AWS IoT, that the flow uses when it executes.

If the definition document doesn't specify a version of the user's namespace, the latest version will be used by default.

Creating a Request

data CreateSystemInstance

See: newCreateSystemInstance smart constructor.

Constructors

CreateSystemInstance'

Fields

metricsConfiguration :: Maybe MetricsConfiguration

greengrassGroupName :: Maybe Text

The name of the Greengrass group where the system instance will be deployed. This value is required if the value of the target parameter is GREENGRASS.

flowActionsRoleArn :: Maybe Text

The ARN of the IAM role that AWS IoT Things Graph will assume when it executes the flow. This role must have read and write access to AWS Lambda and AWS IoT and any other AWS services that the flow uses when it executes. This value is required if the value of the target parameter is CLOUD.

s3BucketName :: Maybe Text

The name of the Amazon Simple Storage Service bucket that will be used to store and deploy the system instance's resource file. This value is required if the value of the target parameter is GREENGRASS.

tags :: Maybe [Tag]

Metadata, consisting of key-value pairs, that can be used to categorize your system instances.

definition :: DefinitionDocument

target :: DeploymentTarget

The target type of the deployment. Valid values are GREENGRASS and CLOUD.

newCreateSystemInstance

Arguments

:: DefinitionDocument
$sel:definition:CreateSystemInstance'

-> DeploymentTarget
$sel:target:CreateSystemInstance'

-> CreateSystemInstance

Create a value of CreateSystemInstance with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:metricsConfiguration:CreateSystemInstance', createSystemInstance_metricsConfiguration - Undocumented member.

$sel:greengrassGroupName:CreateSystemInstance', createSystemInstance_greengrassGroupName - The name of the Greengrass group where the system instance will be deployed. This value is required if the value of the target parameter is GREENGRASS.

$sel:flowActionsRoleArn:CreateSystemInstance', createSystemInstance_flowActionsRoleArn - The ARN of the IAM role that AWS IoT Things Graph will assume when it executes the flow. This role must have read and write access to AWS Lambda and AWS IoT and any other AWS services that the flow uses when it executes. This value is required if the value of the target parameter is CLOUD.

$sel:s3BucketName:CreateSystemInstance', createSystemInstance_s3BucketName - The name of the Amazon Simple Storage Service bucket that will be used to store and deploy the system instance's resource file. This value is required if the value of the target parameter is GREENGRASS.

$sel:tags:CreateSystemInstance', createSystemInstance_tags - Metadata, consisting of key-value pairs, that can be used to categorize your system instances.

$sel:definition:CreateSystemInstance', createSystemInstance_definition - Undocumented member.

$sel:target:CreateSystemInstance', createSystemInstance_target - The target type of the deployment. Valid values are GREENGRASS and CLOUD.

Request Lenses

createSystemInstance_metricsConfiguration :: Lens' CreateSystemInstance (Maybe MetricsConfiguration)

Undocumented member.

createSystemInstance_greengrassGroupName :: Lens' CreateSystemInstance (Maybe Text)

The name of the Greengrass group where the system instance will be deployed. This value is required if the value of the target parameter is GREENGRASS.

createSystemInstance_flowActionsRoleArn :: Lens' CreateSystemInstance (Maybe Text)

The ARN of the IAM role that AWS IoT Things Graph will assume when it executes the flow. This role must have read and write access to AWS Lambda and AWS IoT and any other AWS services that the flow uses when it executes. This value is required if the value of the target parameter is CLOUD.

createSystemInstance_s3BucketName :: Lens' CreateSystemInstance (Maybe Text)

The name of the Amazon Simple Storage Service bucket that will be used to store and deploy the system instance's resource file. This value is required if the value of the target parameter is GREENGRASS.

createSystemInstance_tags :: Lens' CreateSystemInstance (Maybe [Tag])

Metadata, consisting of key-value pairs, that can be used to categorize your system instances.

createSystemInstance_definition :: Lens' CreateSystemInstance DefinitionDocument

Undocumented member.

createSystemInstance_target :: Lens' CreateSystemInstance DeploymentTarget

The target type of the deployment. Valid values are GREENGRASS and CLOUD.

Destructuring the Response

data CreateSystemInstanceResponse

See: newCreateSystemInstanceResponse smart constructor.

Constructors

CreateSystemInstanceResponse'

Fields

summary :: Maybe SystemInstanceSummary

The summary object that describes the new system instance.

httpStatus :: Int

The response's http status code.

newCreateSystemInstanceResponse

Arguments

:: Int
$sel:httpStatus:CreateSystemInstanceResponse'

-> CreateSystemInstanceResponse

Create a value of CreateSystemInstanceResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:summary:CreateSystemInstanceResponse', createSystemInstanceResponse_summary - The summary object that describes the new system instance.

$sel:httpStatus:CreateSystemInstanceResponse', createSystemInstanceResponse_httpStatus - The response's http status code.

Response Lenses

createSystemInstanceResponse_summary :: Lens' CreateSystemInstanceResponse (Maybe SystemInstanceSummary)

The summary object that describes the new system instance.

createSystemInstanceResponse_httpStatus :: Lens' CreateSystemInstanceResponse Int

The response's http status code.