Amazonka.IoTThingsGraph.DeploySystemInstance

Description

Greengrass and Cloud Deployments

Deploys the system instance to the target specified in CreateSystemInstance.

Greengrass Deployments

If the system or any workflows and entities have been updated before this action is called, then the deployment will create a new Amazon Simple Storage Service resource file and then deploy it.

Since this action creates a Greengrass deployment on the caller's behalf, the calling identity must have write permissions to the specified Greengrass group. Otherwise, the call will fail with an authorization error.

For information about the artifacts that get added to your Greengrass core device when you use this API, see AWS IoT Things Graph and AWS IoT Greengrass.

Creating a Request

data DeploySystemInstance

See: newDeploySystemInstance smart constructor.

Constructors

DeploySystemInstance'

Fields

id :: Maybe Text

The ID of the system instance. This value is returned by the CreateSystemInstance action.

The ID should be in the following format.

urn:tdm:REGION/ACCOUNT ID/default:deployment:DEPLOYMENTNAME

newDeploySystemInstance :: DeploySystemInstance

Create a value of DeploySystemInstance with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:id:DeploySystemInstance', deploySystemInstance_id - The ID of the system instance. This value is returned by the CreateSystemInstance action.

The ID should be in the following format.

urn:tdm:REGION/ACCOUNT ID/default:deployment:DEPLOYMENTNAME

Request Lenses

deploySystemInstance_id :: Lens' DeploySystemInstance (Maybe Text)

The ID of the system instance. This value is returned by the CreateSystemInstance action.

The ID should be in the following format.

urn:tdm:REGION/ACCOUNT ID/default:deployment:DEPLOYMENTNAME

Destructuring the Response

data DeploySystemInstanceResponse

See: newDeploySystemInstanceResponse smart constructor.

Constructors

DeploySystemInstanceResponse'

Fields

greengrassDeploymentId :: Maybe Text

The ID of the Greengrass deployment used to deploy the system instance.

httpStatus :: Int

The response's http status code.

summary :: SystemInstanceSummary

An object that contains summary information about a system instance that was deployed.

newDeploySystemInstanceResponse

Arguments

:: Int
$sel:httpStatus:DeploySystemInstanceResponse'

-> SystemInstanceSummary
$sel:summary:DeploySystemInstanceResponse'

-> DeploySystemInstanceResponse

Create a value of DeploySystemInstanceResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:greengrassDeploymentId:DeploySystemInstanceResponse', deploySystemInstanceResponse_greengrassDeploymentId - The ID of the Greengrass deployment used to deploy the system instance.

$sel:httpStatus:DeploySystemInstanceResponse', deploySystemInstanceResponse_httpStatus - The response's http status code.

$sel:summary:DeploySystemInstanceResponse', deploySystemInstanceResponse_summary - An object that contains summary information about a system instance that was deployed.

Response Lenses

deploySystemInstanceResponse_greengrassDeploymentId :: Lens' DeploySystemInstanceResponse (Maybe Text)

The ID of the Greengrass deployment used to deploy the system instance.

deploySystemInstanceResponse_httpStatus :: Lens' DeploySystemInstanceResponse Int

The response's http status code.

deploySystemInstanceResponse_summary :: Lens' DeploySystemInstanceResponse SystemInstanceSummary

An object that contains summary information about a system instance that was deployed.