Amazonka.IoTThingsGraph.GetUploadStatus

Description

Gets the status of the specified upload.

Creating a Request

data GetUploadStatus

See: newGetUploadStatus smart constructor.

Constructors

GetUploadStatus'

Fields

uploadId :: Text

The ID of the upload. This value is returned by the UploadEntityDefinitions action.

newGetUploadStatus

Arguments

:: Text
$sel:uploadId:GetUploadStatus'

-> GetUploadStatus

Create a value of GetUploadStatus with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:uploadId:GetUploadStatus', getUploadStatus_uploadId - The ID of the upload. This value is returned by the UploadEntityDefinitions action.

Request Lenses

getUploadStatus_uploadId :: Lens' GetUploadStatus Text

The ID of the upload. This value is returned by the UploadEntityDefinitions action.

Destructuring the Response

data GetUploadStatusResponse

See: newGetUploadStatusResponse smart constructor.

Constructors

GetUploadStatusResponse'

Fields

failureReason :: Maybe [Text]

The reason for an upload failure.

namespaceArn :: Maybe Text

The ARN of the upload.

namespaceVersion :: Maybe Integer

The version of the user's namespace. Defaults to the latest version of the user's namespace.

namespaceName :: Maybe Text

The name of the upload's namespace.

httpStatus :: Int

The response's http status code.

uploadId :: Text

The ID of the upload.

uploadStatus :: UploadStatus

The status of the upload. The initial status is IN_PROGRESS. The response show all validation failures if the upload fails.

createdDate :: POSIX

The date at which the upload was created.

newGetUploadStatusResponse

Arguments

:: Int
$sel:httpStatus:GetUploadStatusResponse'

-> Text
$sel:uploadId:GetUploadStatus'

-> UploadStatus
$sel:uploadStatus:GetUploadStatusResponse'

-> UTCTime
$sel:createdDate:GetUploadStatusResponse'

-> GetUploadStatusResponse

Create a value of GetUploadStatusResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:failureReason:GetUploadStatusResponse', getUploadStatusResponse_failureReason - The reason for an upload failure.

$sel:namespaceArn:GetUploadStatusResponse', getUploadStatusResponse_namespaceArn - The ARN of the upload.

$sel:namespaceVersion:GetUploadStatusResponse', getUploadStatusResponse_namespaceVersion - The version of the user's namespace. Defaults to the latest version of the user's namespace.

$sel:namespaceName:GetUploadStatusResponse', getUploadStatusResponse_namespaceName - The name of the upload's namespace.

$sel:httpStatus:GetUploadStatusResponse', getUploadStatusResponse_httpStatus - The response's http status code.

$sel:uploadId:GetUploadStatus', getUploadStatusResponse_uploadId - The ID of the upload.

$sel:uploadStatus:GetUploadStatusResponse', getUploadStatusResponse_uploadStatus - The status of the upload. The initial status is IN_PROGRESS. The response show all validation failures if the upload fails.

$sel:createdDate:GetUploadStatusResponse', getUploadStatusResponse_createdDate - The date at which the upload was created.

Response Lenses

getUploadStatusResponse_failureReason :: Lens' GetUploadStatusResponse (Maybe [Text])

The reason for an upload failure.

getUploadStatusResponse_namespaceArn :: Lens' GetUploadStatusResponse (Maybe Text)

The ARN of the upload.

getUploadStatusResponse_namespaceVersion :: Lens' GetUploadStatusResponse (Maybe Integer)

The version of the user's namespace. Defaults to the latest version of the user's namespace.

getUploadStatusResponse_namespaceName :: Lens' GetUploadStatusResponse (Maybe Text)

The name of the upload's namespace.

getUploadStatusResponse_httpStatus :: Lens' GetUploadStatusResponse Int

The response's http status code.

getUploadStatusResponse_uploadId :: Lens' GetUploadStatusResponse Text

The ID of the upload.

getUploadStatusResponse_uploadStatus :: Lens' GetUploadStatusResponse UploadStatus

The status of the upload. The initial status is IN_PROGRESS. The response show all validation failures if the upload fails.

getUploadStatusResponse_createdDate :: Lens' GetUploadStatusResponse UTCTime

The date at which the upload was created.