Amazonka.IoTThingsGraph.Types

Description

Service Configuration

defaultService :: Service

API version 2018-09-06 of the Amazon IoT Things Graph SDK configuration.

Errors

_InvalidRequestException :: AsError a => Getting (First ServiceError) a ServiceError

_ResourceAlreadyExistsException :: AsError a => Getting (First ServiceError) a ServiceError

_ThrottlingException :: AsError a => Getting (First ServiceError) a ServiceError

_InternalFailureException :: AsError a => Getting (First ServiceError) a ServiceError

_ResourceNotFoundException :: AsError a => Getting (First ServiceError) a ServiceError

_LimitExceededException :: AsError a => Getting (First ServiceError) a ServiceError

_ResourceInUseException :: AsError a => Getting (First ServiceError) a ServiceError

DefinitionLanguage

newtype DefinitionLanguage

Constructors

DefinitionLanguage'

Fields

fromDefinitionLanguage :: Text

Bundled Patterns

pattern DefinitionLanguage_GRAPHQL :: DefinitionLanguage

DeploymentTarget

newtype DeploymentTarget

Constructors

DeploymentTarget'

Fields

fromDeploymentTarget :: Text

Bundled Patterns

pattern DeploymentTarget_CLOUD :: DeploymentTarget

pattern DeploymentTarget_GREENGRASS :: DeploymentTarget

EntityFilterName

newtype EntityFilterName

Constructors

EntityFilterName'

Fields

fromEntityFilterName :: Text

Bundled Patterns

pattern EntityFilterName_NAME :: EntityFilterName

pattern EntityFilterName_NAMESPACE :: EntityFilterName

pattern EntityFilterName_REFERENCED_ENTITY_ID :: EntityFilterName

pattern EntityFilterName_SEMANTIC_TYPE_PATH :: EntityFilterName

EntityType

newtype EntityType

Constructors

EntityType'

Fields

fromEntityType :: Text

Bundled Patterns

pattern EntityType_ACTION :: EntityType

pattern EntityType_CAPABILITY :: EntityType

pattern EntityType_DEVICE :: EntityType

pattern EntityType_DEVICE_MODEL :: EntityType

pattern EntityType_ENUM :: EntityType

pattern EntityType_EVENT :: EntityType

pattern EntityType_MAPPING :: EntityType

pattern EntityType_PROPERTY :: EntityType

pattern EntityType_SERVICE :: EntityType

pattern EntityType_STATE :: EntityType

FlowExecutionEventType

newtype FlowExecutionEventType

Constructors

FlowExecutionEventType'

Fields

fromFlowExecutionEventType :: Text

Bundled Patterns

pattern FlowExecutionEventType_ACKNOWLEDGE_TASK_MESSAGE :: FlowExecutionEventType

pattern FlowExecutionEventType_ACTIVITY_FAILED :: FlowExecutionEventType

pattern FlowExecutionEventType_ACTIVITY_SCHEDULED :: FlowExecutionEventType

pattern FlowExecutionEventType_ACTIVITY_STARTED :: FlowExecutionEventType

pattern FlowExecutionEventType_ACTIVITY_SUCCEEDED :: FlowExecutionEventType

pattern FlowExecutionEventType_EXECUTION_ABORTED :: FlowExecutionEventType

pattern FlowExecutionEventType_EXECUTION_FAILED :: FlowExecutionEventType

pattern FlowExecutionEventType_EXECUTION_STARTED :: FlowExecutionEventType

pattern FlowExecutionEventType_EXECUTION_SUCCEEDED :: FlowExecutionEventType

pattern FlowExecutionEventType_SCHEDULE_NEXT_READY_STEPS_TASK :: FlowExecutionEventType

pattern FlowExecutionEventType_START_FLOW_EXECUTION_TASK :: FlowExecutionEventType

pattern FlowExecutionEventType_STEP_FAILED :: FlowExecutionEventType

pattern FlowExecutionEventType_STEP_STARTED :: FlowExecutionEventType

pattern FlowExecutionEventType_STEP_SUCCEEDED :: FlowExecutionEventType

pattern FlowExecutionEventType_THING_ACTION_TASK :: FlowExecutionEventType

pattern FlowExecutionEventType_THING_ACTION_TASK_FAILED :: FlowExecutionEventType

pattern FlowExecutionEventType_THING_ACTION_TASK_SUCCEEDED :: FlowExecutionEventType

FlowExecutionStatus

newtype FlowExecutionStatus

Constructors

FlowExecutionStatus'

Fields

fromFlowExecutionStatus :: Text

Bundled Patterns

pattern FlowExecutionStatus_ABORTED :: FlowExecutionStatus

pattern FlowExecutionStatus_FAILED :: FlowExecutionStatus

pattern FlowExecutionStatus_RUNNING :: FlowExecutionStatus

pattern FlowExecutionStatus_SUCCEEDED :: FlowExecutionStatus

FlowTemplateFilterName

newtype FlowTemplateFilterName

Constructors

FlowTemplateFilterName'

Fields

fromFlowTemplateFilterName :: Text

Bundled Patterns

pattern FlowTemplateFilterName_DEVICE_MODEL_ID :: FlowTemplateFilterName

NamespaceDeletionStatus

newtype NamespaceDeletionStatus

Constructors

NamespaceDeletionStatus'

Fields

fromNamespaceDeletionStatus :: Text

Bundled Patterns

pattern NamespaceDeletionStatus_FAILED :: NamespaceDeletionStatus

pattern NamespaceDeletionStatus_IN_PROGRESS :: NamespaceDeletionStatus

pattern NamespaceDeletionStatus_SUCCEEDED :: NamespaceDeletionStatus

NamespaceDeletionStatusErrorCodes

newtype NamespaceDeletionStatusErrorCodes

Constructors

NamespaceDeletionStatusErrorCodes'

Fields

fromNamespaceDeletionStatusErrorCodes :: Text

Bundled Patterns

pattern NamespaceDeletionStatusErrorCodes_VALIDATION_FAILED :: NamespaceDeletionStatusErrorCodes

SystemInstanceDeploymentStatus

newtype SystemInstanceDeploymentStatus

Constructors

SystemInstanceDeploymentStatus'

Fields

fromSystemInstanceDeploymentStatus :: Text

Bundled Patterns

pattern SystemInstanceDeploymentStatus_BOOTSTRAP :: SystemInstanceDeploymentStatus

pattern SystemInstanceDeploymentStatus_DELETED_IN_TARGET :: SystemInstanceDeploymentStatus

pattern SystemInstanceDeploymentStatus_DEPLOYED_IN_TARGET :: SystemInstanceDeploymentStatus

pattern SystemInstanceDeploymentStatus_DEPLOY_IN_PROGRESS :: SystemInstanceDeploymentStatus

pattern SystemInstanceDeploymentStatus_FAILED :: SystemInstanceDeploymentStatus

pattern SystemInstanceDeploymentStatus_NOT_DEPLOYED :: SystemInstanceDeploymentStatus

pattern SystemInstanceDeploymentStatus_PENDING_DELETE :: SystemInstanceDeploymentStatus

pattern SystemInstanceDeploymentStatus_UNDEPLOY_IN_PROGRESS :: SystemInstanceDeploymentStatus

SystemInstanceFilterName

newtype SystemInstanceFilterName

Constructors

SystemInstanceFilterName'

Fields

fromSystemInstanceFilterName :: Text

Bundled Patterns

pattern SystemInstanceFilterName_GREENGRASS_GROUP_NAME :: SystemInstanceFilterName

pattern SystemInstanceFilterName_STATUS :: SystemInstanceFilterName

pattern SystemInstanceFilterName_SYSTEM_TEMPLATE_ID :: SystemInstanceFilterName

SystemTemplateFilterName

newtype SystemTemplateFilterName

Constructors

SystemTemplateFilterName'

Fields

fromSystemTemplateFilterName :: Text

Bundled Patterns

pattern SystemTemplateFilterName_FLOW_TEMPLATE_ID :: SystemTemplateFilterName

UploadStatus

newtype UploadStatus

Constructors

UploadStatus'

Fields

fromUploadStatus :: Text

Bundled Patterns

pattern UploadStatus_FAILED :: UploadStatus

pattern UploadStatus_IN_PROGRESS :: UploadStatus

pattern UploadStatus_SUCCEEDED :: UploadStatus

DefinitionDocument

data DefinitionDocument

A document that defines an entity.

See: newDefinitionDocument smart constructor.

Constructors

DefinitionDocument'

Fields

language :: DefinitionLanguage

The language used to define the entity. GRAPHQL is the only valid value.

text :: Text

The GraphQL text that defines the entity.

newDefinitionDocument

Arguments

:: DefinitionLanguage
$sel:language:DefinitionDocument'

-> Text
$sel:text:DefinitionDocument'

-> DefinitionDocument

Create a value of DefinitionDocument with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:language:DefinitionDocument', definitionDocument_language - The language used to define the entity. GRAPHQL is the only valid value.

$sel:text:DefinitionDocument', definitionDocument_text - The GraphQL text that defines the entity.

definitionDocument_language :: Lens' DefinitionDocument DefinitionLanguage

The language used to define the entity. GRAPHQL is the only valid value.

definitionDocument_text :: Lens' DefinitionDocument Text

The GraphQL text that defines the entity.

DependencyRevision

data DependencyRevision

An object that contains the ID and revision number of a workflow or system that is part of a deployment.

See: newDependencyRevision smart constructor.

Constructors

DependencyRevision'

Fields

revisionNumber :: Maybe Integer

The revision number of the workflow or system.

id :: Maybe Text

The ID of the workflow or system.

newDependencyRevision :: DependencyRevision

Create a value of DependencyRevision with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:revisionNumber:DependencyRevision', dependencyRevision_revisionNumber - The revision number of the workflow or system.

$sel:id:DependencyRevision', dependencyRevision_id - The ID of the workflow or system.

dependencyRevision_revisionNumber :: Lens' DependencyRevision (Maybe Integer)

The revision number of the workflow or system.

dependencyRevision_id :: Lens' DependencyRevision (Maybe Text)

The ID of the workflow or system.

EntityDescription

data EntityDescription

Describes the properties of an entity.

See: newEntityDescription smart constructor.

Constructors

EntityDescription'

Fields

arn :: Maybe Text

The entity ARN.

createdAt :: Maybe POSIX

The time at which the entity was created.

definition :: Maybe DefinitionDocument

The definition document of the entity.

id :: Maybe Text

The entity ID.

type' :: Maybe EntityType

The entity type.

newEntityDescription :: EntityDescription

Create a value of EntityDescription with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:arn:EntityDescription', entityDescription_arn - The entity ARN.

$sel:createdAt:EntityDescription', entityDescription_createdAt - The time at which the entity was created.

$sel:definition:EntityDescription', entityDescription_definition - The definition document of the entity.

$sel:id:EntityDescription', entityDescription_id - The entity ID.

$sel:type':EntityDescription', entityDescription_type - The entity type.

entityDescription_arn :: Lens' EntityDescription (Maybe Text)

The entity ARN.

entityDescription_createdAt :: Lens' EntityDescription (Maybe UTCTime)

The time at which the entity was created.

entityDescription_definition :: Lens' EntityDescription (Maybe DefinitionDocument)

The definition document of the entity.

entityDescription_id :: Lens' EntityDescription (Maybe Text)

The entity ID.

entityDescription_type :: Lens' EntityDescription (Maybe EntityType)

The entity type.

EntityFilter

data EntityFilter

An object that filters an entity search. Multiple filters function as OR criteria in the search. For example a search that includes a NAMESPACE and a REFERENCED_ENTITY_ID filter searches for entities in the specified namespace that use the entity specified by the value of REFERENCED_ENTITY_ID.

See: newEntityFilter smart constructor.

Constructors

EntityFilter'

Fields

value :: Maybe [Text]

An array of string values for the search filter field. Multiple values function as AND criteria in the search.

name :: Maybe EntityFilterName

The name of the entity search filter field. REFERENCED_ENTITY_ID filters on entities that are used by the entity in the result set. For example, you can filter on the ID of a property that is used in a state.

newEntityFilter :: EntityFilter

Create a value of EntityFilter with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:value:EntityFilter', entityFilter_value - An array of string values for the search filter field. Multiple values function as AND criteria in the search.

$sel:name:EntityFilter', entityFilter_name - The name of the entity search filter field. REFERENCED_ENTITY_ID filters on entities that are used by the entity in the result set. For example, you can filter on the ID of a property that is used in a state.

entityFilter_value :: Lens' EntityFilter (Maybe [Text])

An array of string values for the search filter field. Multiple values function as AND criteria in the search.

entityFilter_name :: Lens' EntityFilter (Maybe EntityFilterName)

The name of the entity search filter field. REFERENCED_ENTITY_ID filters on entities that are used by the entity in the result set. For example, you can filter on the ID of a property that is used in a state.

FlowExecutionMessage

data FlowExecutionMessage

An object that contains information about a flow event.

See: newFlowExecutionMessage smart constructor.

Constructors

FlowExecutionMessage'

Fields

payload :: Maybe Text

A string containing information about the flow event.

eventType :: Maybe FlowExecutionEventType

The type of flow event .

timestamp :: Maybe POSIX

The date and time when the message was last updated.

messageId :: Maybe Text

The unique identifier of the message.

newFlowExecutionMessage :: FlowExecutionMessage

Create a value of FlowExecutionMessage with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:payload:FlowExecutionMessage', flowExecutionMessage_payload - A string containing information about the flow event.

$sel:eventType:FlowExecutionMessage', flowExecutionMessage_eventType - The type of flow event .

$sel:timestamp:FlowExecutionMessage', flowExecutionMessage_timestamp - The date and time when the message was last updated.

$sel:messageId:FlowExecutionMessage', flowExecutionMessage_messageId - The unique identifier of the message.

flowExecutionMessage_payload :: Lens' FlowExecutionMessage (Maybe Text)

A string containing information about the flow event.

flowExecutionMessage_eventType :: Lens' FlowExecutionMessage (Maybe FlowExecutionEventType)

The type of flow event .

flowExecutionMessage_timestamp :: Lens' FlowExecutionMessage (Maybe UTCTime)

The date and time when the message was last updated.

flowExecutionMessage_messageId :: Lens' FlowExecutionMessage (Maybe Text)

The unique identifier of the message.

FlowExecutionSummary

data FlowExecutionSummary

An object that contains summary information about a flow execution.

See: newFlowExecutionSummary smart constructor.

Constructors

FlowExecutionSummary'

Fields

status :: Maybe FlowExecutionStatus

The current status of the flow execution.

flowTemplateId :: Maybe Text

The ID of the flow.

createdAt :: Maybe POSIX

The date and time when the flow execution summary was created.

flowExecutionId :: Maybe Text

The ID of the flow execution.

systemInstanceId :: Maybe Text

The ID of the system instance that contains the flow.

updatedAt :: Maybe POSIX

The date and time when the flow execution summary was last updated.

newFlowExecutionSummary :: FlowExecutionSummary

Create a value of FlowExecutionSummary with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:status:FlowExecutionSummary', flowExecutionSummary_status - The current status of the flow execution.

$sel:flowTemplateId:FlowExecutionSummary', flowExecutionSummary_flowTemplateId - The ID of the flow.

$sel:createdAt:FlowExecutionSummary', flowExecutionSummary_createdAt - The date and time when the flow execution summary was created.

$sel:flowExecutionId:FlowExecutionSummary', flowExecutionSummary_flowExecutionId - The ID of the flow execution.

$sel:systemInstanceId:FlowExecutionSummary', flowExecutionSummary_systemInstanceId - The ID of the system instance that contains the flow.

$sel:updatedAt:FlowExecutionSummary', flowExecutionSummary_updatedAt - The date and time when the flow execution summary was last updated.

flowExecutionSummary_status :: Lens' FlowExecutionSummary (Maybe FlowExecutionStatus)

The current status of the flow execution.

flowExecutionSummary_flowTemplateId :: Lens' FlowExecutionSummary (Maybe Text)

The ID of the flow.

flowExecutionSummary_createdAt :: Lens' FlowExecutionSummary (Maybe UTCTime)

The date and time when the flow execution summary was created.

flowExecutionSummary_flowExecutionId :: Lens' FlowExecutionSummary (Maybe Text)

The ID of the flow execution.

flowExecutionSummary_systemInstanceId :: Lens' FlowExecutionSummary (Maybe Text)

The ID of the system instance that contains the flow.

flowExecutionSummary_updatedAt :: Lens' FlowExecutionSummary (Maybe UTCTime)

The date and time when the flow execution summary was last updated.

FlowTemplateDescription

data FlowTemplateDescription

An object that contains a workflow's definition and summary information.

See: newFlowTemplateDescription smart constructor.

Constructors

FlowTemplateDescription'

Fields

summary :: Maybe FlowTemplateSummary

An object that contains summary information about a workflow.

definition :: Maybe DefinitionDocument

A workflow's definition document.

validatedNamespaceVersion :: Maybe Integer

The version of the user's namespace against which the workflow was validated. Use this value in your system instance.

newFlowTemplateDescription :: FlowTemplateDescription

Create a value of FlowTemplateDescription with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:summary:FlowTemplateDescription', flowTemplateDescription_summary - An object that contains summary information about a workflow.

$sel:definition:FlowTemplateDescription', flowTemplateDescription_definition - A workflow's definition document.

$sel:validatedNamespaceVersion:FlowTemplateDescription', flowTemplateDescription_validatedNamespaceVersion - The version of the user's namespace against which the workflow was validated. Use this value in your system instance.

flowTemplateDescription_summary :: Lens' FlowTemplateDescription (Maybe FlowTemplateSummary)

An object that contains summary information about a workflow.

flowTemplateDescription_definition :: Lens' FlowTemplateDescription (Maybe DefinitionDocument)

A workflow's definition document.

flowTemplateDescription_validatedNamespaceVersion :: Lens' FlowTemplateDescription (Maybe Integer)

The version of the user's namespace against which the workflow was validated. Use this value in your system instance.

FlowTemplateFilter

data FlowTemplateFilter

An object that filters a workflow search.

See: newFlowTemplateFilter smart constructor.

Constructors

FlowTemplateFilter'

Fields

name :: FlowTemplateFilterName

The name of the search filter field.

value :: [Text]

An array of string values for the search filter field. Multiple values function as AND criteria in the search.

newFlowTemplateFilter

Arguments

:: FlowTemplateFilterName
$sel:name:FlowTemplateFilter'

-> FlowTemplateFilter

Create a value of FlowTemplateFilter with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:name:FlowTemplateFilter', flowTemplateFilter_name - The name of the search filter field.

$sel:value:FlowTemplateFilter', flowTemplateFilter_value - An array of string values for the search filter field. Multiple values function as AND criteria in the search.

flowTemplateFilter_name :: Lens' FlowTemplateFilter FlowTemplateFilterName

The name of the search filter field.

flowTemplateFilter_value :: Lens' FlowTemplateFilter [Text]

An array of string values for the search filter field. Multiple values function as AND criteria in the search.

FlowTemplateSummary

data FlowTemplateSummary

An object that contains summary information about a workflow.

See: newFlowTemplateSummary smart constructor.

Constructors

FlowTemplateSummary'

Fields

arn :: Maybe Text

The ARN of the workflow.

createdAt :: Maybe POSIX

The date when the workflow was created.

revisionNumber :: Maybe Integer

The revision number of the workflow.

id :: Maybe Text

The ID of the workflow.

newFlowTemplateSummary :: FlowTemplateSummary

Create a value of FlowTemplateSummary with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:arn:FlowTemplateSummary', flowTemplateSummary_arn - The ARN of the workflow.

$sel:createdAt:FlowTemplateSummary', flowTemplateSummary_createdAt - The date when the workflow was created.

$sel:revisionNumber:FlowTemplateSummary', flowTemplateSummary_revisionNumber - The revision number of the workflow.

$sel:id:FlowTemplateSummary', flowTemplateSummary_id - The ID of the workflow.

flowTemplateSummary_arn :: Lens' FlowTemplateSummary (Maybe Text)

The ARN of the workflow.

flowTemplateSummary_createdAt :: Lens' FlowTemplateSummary (Maybe UTCTime)

The date when the workflow was created.

flowTemplateSummary_revisionNumber :: Lens' FlowTemplateSummary (Maybe Integer)

The revision number of the workflow.

flowTemplateSummary_id :: Lens' FlowTemplateSummary (Maybe Text)

The ID of the workflow.

MetricsConfiguration

data MetricsConfiguration

An object that specifies whether cloud metrics are collected in a deployment and, if so, what role is used to collect metrics.

See: newMetricsConfiguration smart constructor.

Constructors

MetricsConfiguration'

Fields

cloudMetricEnabled :: Maybe Bool

A Boolean that specifies whether cloud metrics are collected.

metricRuleRoleArn :: Maybe Text

The ARN of the role that is used to collect cloud metrics.

newMetricsConfiguration :: MetricsConfiguration

Create a value of MetricsConfiguration with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:cloudMetricEnabled:MetricsConfiguration', metricsConfiguration_cloudMetricEnabled - A Boolean that specifies whether cloud metrics are collected.

$sel:metricRuleRoleArn:MetricsConfiguration', metricsConfiguration_metricRuleRoleArn - The ARN of the role that is used to collect cloud metrics.

metricsConfiguration_cloudMetricEnabled :: Lens' MetricsConfiguration (Maybe Bool)

A Boolean that specifies whether cloud metrics are collected.

metricsConfiguration_metricRuleRoleArn :: Lens' MetricsConfiguration (Maybe Text)

The ARN of the role that is used to collect cloud metrics.

SystemInstanceDescription

data SystemInstanceDescription

An object that contains a system instance definition and summary information.

See: newSystemInstanceDescription smart constructor.

Constructors

SystemInstanceDescription'

Fields

summary :: Maybe SystemInstanceSummary

An object that contains summary information about a system instance.

metricsConfiguration :: Maybe MetricsConfiguration

validatedDependencyRevisions :: Maybe [DependencyRevision]

A list of objects that contain all of the IDs and revision numbers of workflows and systems that are used in a system instance.

definition :: Maybe DefinitionDocument

validatedNamespaceVersion :: Maybe Integer

The version of the user's namespace against which the system instance was validated.

flowActionsRoleArn :: Maybe Text

The AWS Identity and Access Management (IAM) role that AWS IoT Things Graph assumes during flow execution in a cloud deployment. This role must have read and write permissionss to AWS Lambda and AWS IoT and to any other AWS services that the flow uses.

s3BucketName :: Maybe Text

The Amazon Simple Storage Service bucket where information about a system instance is stored.

newSystemInstanceDescription :: SystemInstanceDescription

Create a value of SystemInstanceDescription with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:summary:SystemInstanceDescription', systemInstanceDescription_summary - An object that contains summary information about a system instance.

$sel:metricsConfiguration:SystemInstanceDescription', systemInstanceDescription_metricsConfiguration - Undocumented member.

$sel:validatedDependencyRevisions:SystemInstanceDescription', systemInstanceDescription_validatedDependencyRevisions - A list of objects that contain all of the IDs and revision numbers of workflows and systems that are used in a system instance.

$sel:definition:SystemInstanceDescription', systemInstanceDescription_definition - Undocumented member.

$sel:validatedNamespaceVersion:SystemInstanceDescription', systemInstanceDescription_validatedNamespaceVersion - The version of the user's namespace against which the system instance was validated.

$sel:flowActionsRoleArn:SystemInstanceDescription', systemInstanceDescription_flowActionsRoleArn - The AWS Identity and Access Management (IAM) role that AWS IoT Things Graph assumes during flow execution in a cloud deployment. This role must have read and write permissionss to AWS Lambda and AWS IoT and to any other AWS services that the flow uses.

$sel:s3BucketName:SystemInstanceDescription', systemInstanceDescription_s3BucketName - The Amazon Simple Storage Service bucket where information about a system instance is stored.

systemInstanceDescription_summary :: Lens' SystemInstanceDescription (Maybe SystemInstanceSummary)

An object that contains summary information about a system instance.

systemInstanceDescription_metricsConfiguration :: Lens' SystemInstanceDescription (Maybe MetricsConfiguration)

Undocumented member.

systemInstanceDescription_validatedDependencyRevisions :: Lens' SystemInstanceDescription (Maybe [DependencyRevision])

A list of objects that contain all of the IDs and revision numbers of workflows and systems that are used in a system instance.

systemInstanceDescription_definition :: Lens' SystemInstanceDescription (Maybe DefinitionDocument)

Undocumented member.

systemInstanceDescription_validatedNamespaceVersion :: Lens' SystemInstanceDescription (Maybe Integer)

The version of the user's namespace against which the system instance was validated.

systemInstanceDescription_flowActionsRoleArn :: Lens' SystemInstanceDescription (Maybe Text)

The AWS Identity and Access Management (IAM) role that AWS IoT Things Graph assumes during flow execution in a cloud deployment. This role must have read and write permissionss to AWS Lambda and AWS IoT and to any other AWS services that the flow uses.

systemInstanceDescription_s3BucketName :: Lens' SystemInstanceDescription (Maybe Text)

The Amazon Simple Storage Service bucket where information about a system instance is stored.

SystemInstanceFilter

data SystemInstanceFilter

An object that filters a system instance search. Multiple filters function as OR criteria in the search. For example a search that includes a GREENGRASS_GROUP_NAME and a STATUS filter searches for system instances in the specified Greengrass group that have the specified status.

See: newSystemInstanceFilter smart constructor.

Constructors

SystemInstanceFilter'

Fields

value :: Maybe [Text]

An array of string values for the search filter field. Multiple values function as AND criteria in the search.

name :: Maybe SystemInstanceFilterName

The name of the search filter field.

newSystemInstanceFilter :: SystemInstanceFilter

Create a value of SystemInstanceFilter with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:value:SystemInstanceFilter', systemInstanceFilter_value - An array of string values for the search filter field. Multiple values function as AND criteria in the search.

$sel:name:SystemInstanceFilter', systemInstanceFilter_name - The name of the search filter field.

systemInstanceFilter_value :: Lens' SystemInstanceFilter (Maybe [Text])

An array of string values for the search filter field. Multiple values function as AND criteria in the search.

systemInstanceFilter_name :: Lens' SystemInstanceFilter (Maybe SystemInstanceFilterName)

The name of the search filter field.

SystemInstanceSummary

data SystemInstanceSummary

An object that contains summary information about a system instance.

See: newSystemInstanceSummary smart constructor.

Constructors

SystemInstanceSummary'

Fields

status :: Maybe SystemInstanceDeploymentStatus

The status of the system instance.

greengrassGroupName :: Maybe Text

The ID of the Greengrass group where the system instance is deployed.

arn :: Maybe Text

The ARN of the system instance.

createdAt :: Maybe POSIX

The date when the system instance was created.

greengrassGroupId :: Maybe Text

The ID of the Greengrass group where the system instance is deployed.

greengrassGroupVersionId :: Maybe Text

The version of the Greengrass group where the system instance is deployed.

id :: Maybe Text

The ID of the system instance.

updatedAt :: Maybe POSIX

The date and time when the system instance was last updated.

target :: Maybe DeploymentTarget

The target of the system instance.

newSystemInstanceSummary :: SystemInstanceSummary

Create a value of SystemInstanceSummary with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:status:SystemInstanceSummary', systemInstanceSummary_status - The status of the system instance.

$sel:greengrassGroupName:SystemInstanceSummary', systemInstanceSummary_greengrassGroupName - The ID of the Greengrass group where the system instance is deployed.

$sel:arn:SystemInstanceSummary', systemInstanceSummary_arn - The ARN of the system instance.

$sel:createdAt:SystemInstanceSummary', systemInstanceSummary_createdAt - The date when the system instance was created.

$sel:greengrassGroupId:SystemInstanceSummary', systemInstanceSummary_greengrassGroupId - The ID of the Greengrass group where the system instance is deployed.

$sel:greengrassGroupVersionId:SystemInstanceSummary', systemInstanceSummary_greengrassGroupVersionId - The version of the Greengrass group where the system instance is deployed.

$sel:id:SystemInstanceSummary', systemInstanceSummary_id - The ID of the system instance.

$sel:updatedAt:SystemInstanceSummary', systemInstanceSummary_updatedAt - The date and time when the system instance was last updated.

$sel:target:SystemInstanceSummary', systemInstanceSummary_target - The target of the system instance.

systemInstanceSummary_status :: Lens' SystemInstanceSummary (Maybe SystemInstanceDeploymentStatus)

The status of the system instance.

systemInstanceSummary_greengrassGroupName :: Lens' SystemInstanceSummary (Maybe Text)

The ID of the Greengrass group where the system instance is deployed.

systemInstanceSummary_arn :: Lens' SystemInstanceSummary (Maybe Text)

The ARN of the system instance.

systemInstanceSummary_createdAt :: Lens' SystemInstanceSummary (Maybe UTCTime)

The date when the system instance was created.

systemInstanceSummary_greengrassGroupId :: Lens' SystemInstanceSummary (Maybe Text)

The ID of the Greengrass group where the system instance is deployed.

systemInstanceSummary_greengrassGroupVersionId :: Lens' SystemInstanceSummary (Maybe Text)

The version of the Greengrass group where the system instance is deployed.

systemInstanceSummary_id :: Lens' SystemInstanceSummary (Maybe Text)

The ID of the system instance.

systemInstanceSummary_updatedAt :: Lens' SystemInstanceSummary (Maybe UTCTime)

The date and time when the system instance was last updated.

systemInstanceSummary_target :: Lens' SystemInstanceSummary (Maybe DeploymentTarget)

The target of the system instance.

SystemTemplateDescription

data SystemTemplateDescription

An object that contains a system's definition document and summary information.

See: newSystemTemplateDescription smart constructor.

Constructors

SystemTemplateDescription'

Fields

summary :: Maybe SystemTemplateSummary

An object that contains summary information about a system.

definition :: Maybe DefinitionDocument

The definition document of a system.

validatedNamespaceVersion :: Maybe Integer

The namespace version against which the system was validated. Use this value in your system instance.

newSystemTemplateDescription :: SystemTemplateDescription

Create a value of SystemTemplateDescription with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:summary:SystemTemplateDescription', systemTemplateDescription_summary - An object that contains summary information about a system.

$sel:definition:SystemTemplateDescription', systemTemplateDescription_definition - The definition document of a system.

$sel:validatedNamespaceVersion:SystemTemplateDescription', systemTemplateDescription_validatedNamespaceVersion - The namespace version against which the system was validated. Use this value in your system instance.

systemTemplateDescription_summary :: Lens' SystemTemplateDescription (Maybe SystemTemplateSummary)

An object that contains summary information about a system.

systemTemplateDescription_definition :: Lens' SystemTemplateDescription (Maybe DefinitionDocument)

The definition document of a system.

systemTemplateDescription_validatedNamespaceVersion :: Lens' SystemTemplateDescription (Maybe Integer)

The namespace version against which the system was validated. Use this value in your system instance.

SystemTemplateFilter

data SystemTemplateFilter

An object that filters a system search.

See: newSystemTemplateFilter smart constructor.

Constructors

SystemTemplateFilter'

Fields

name :: SystemTemplateFilterName

The name of the system search filter field.

value :: [Text]

An array of string values for the search filter field. Multiple values function as AND criteria in the search.

newSystemTemplateFilter

Arguments

:: SystemTemplateFilterName
$sel:name:SystemTemplateFilter'

-> SystemTemplateFilter

Create a value of SystemTemplateFilter with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:name:SystemTemplateFilter', systemTemplateFilter_name - The name of the system search filter field.

$sel:value:SystemTemplateFilter', systemTemplateFilter_value - An array of string values for the search filter field. Multiple values function as AND criteria in the search.

systemTemplateFilter_name :: Lens' SystemTemplateFilter SystemTemplateFilterName

The name of the system search filter field.

systemTemplateFilter_value :: Lens' SystemTemplateFilter [Text]

An array of string values for the search filter field. Multiple values function as AND criteria in the search.

SystemTemplateSummary

data SystemTemplateSummary

An object that contains information about a system.

See: newSystemTemplateSummary smart constructor.

Constructors

SystemTemplateSummary'

Fields

arn :: Maybe Text

The ARN of the system.

createdAt :: Maybe POSIX

The date when the system was created.

revisionNumber :: Maybe Integer

The revision number of the system.

id :: Maybe Text

The ID of the system.

newSystemTemplateSummary :: SystemTemplateSummary

Create a value of SystemTemplateSummary with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:arn:SystemTemplateSummary', systemTemplateSummary_arn - The ARN of the system.

$sel:createdAt:SystemTemplateSummary', systemTemplateSummary_createdAt - The date when the system was created.

$sel:revisionNumber:SystemTemplateSummary', systemTemplateSummary_revisionNumber - The revision number of the system.

$sel:id:SystemTemplateSummary', systemTemplateSummary_id - The ID of the system.

systemTemplateSummary_arn :: Lens' SystemTemplateSummary (Maybe Text)

The ARN of the system.

systemTemplateSummary_createdAt :: Lens' SystemTemplateSummary (Maybe UTCTime)

The date when the system was created.

systemTemplateSummary_revisionNumber :: Lens' SystemTemplateSummary (Maybe Integer)

The revision number of the system.

systemTemplateSummary_id :: Lens' SystemTemplateSummary (Maybe Text)

The ID of the system.

Tag

data Tag

Metadata assigned to an AWS IoT Things Graph resource consisting of a key-value pair.

See: newTag smart constructor.

Constructors

Tag'

Fields

key :: Text

The required name of the tag. The string value can be from 1 to 128 Unicode characters in length.

value :: Text

The optional value of the tag. The string value can be from 1 to 256 Unicode characters in length.

newTag

Arguments

:: Text
$sel:key:Tag'

-> Text
$sel:value:Tag'

-> Tag

Create a value of Tag with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:key:Tag', tag_key - The required name of the tag. The string value can be from 1 to 128 Unicode characters in length.

$sel:value:Tag', tag_value - The optional value of the tag. The string value can be from 1 to 256 Unicode characters in length.

tag_key :: Lens' Tag Text

The required name of the tag. The string value can be from 1 to 128 Unicode characters in length.

tag_value :: Lens' Tag Text

The optional value of the tag. The string value can be from 1 to 256 Unicode characters in length.

Thing

data Thing

An AWS IoT thing.

See: newThing smart constructor.

Constructors

Thing'

Fields

thingArn :: Maybe Text

The ARN of the thing.

thingName :: Maybe Text

The name of the thing.

newThing :: Thing

Create a value of Thing with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:thingArn:Thing', thing_thingArn - The ARN of the thing.

$sel:thingName:Thing', thing_thingName - The name of the thing.

thing_thingArn :: Lens' Thing (Maybe Text)

The ARN of the thing.

thing_thingName :: Lens' Thing (Maybe Text)

The name of the thing.