Amazonka.IoTThingsGraph.Types.FlowExecutionSummary

Description

Documentation

data FlowExecutionSummary

An object that contains summary information about a flow execution.

See: newFlowExecutionSummary smart constructor.

Constructors

FlowExecutionSummary'

Fields

status :: Maybe FlowExecutionStatus

The current status of the flow execution.

flowTemplateId :: Maybe Text

The ID of the flow.

createdAt :: Maybe POSIX

The date and time when the flow execution summary was created.

flowExecutionId :: Maybe Text

The ID of the flow execution.

systemInstanceId :: Maybe Text

The ID of the system instance that contains the flow.

updatedAt :: Maybe POSIX

The date and time when the flow execution summary was last updated.

newFlowExecutionSummary :: FlowExecutionSummary

Create a value of FlowExecutionSummary with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:status:FlowExecutionSummary', flowExecutionSummary_status - The current status of the flow execution.

$sel:flowTemplateId:FlowExecutionSummary', flowExecutionSummary_flowTemplateId - The ID of the flow.

$sel:createdAt:FlowExecutionSummary', flowExecutionSummary_createdAt - The date and time when the flow execution summary was created.

$sel:flowExecutionId:FlowExecutionSummary', flowExecutionSummary_flowExecutionId - The ID of the flow execution.

$sel:systemInstanceId:FlowExecutionSummary', flowExecutionSummary_systemInstanceId - The ID of the system instance that contains the flow.

$sel:updatedAt:FlowExecutionSummary', flowExecutionSummary_updatedAt - The date and time when the flow execution summary was last updated.

flowExecutionSummary_status :: Lens' FlowExecutionSummary (Maybe FlowExecutionStatus)

The current status of the flow execution.

flowExecutionSummary_flowTemplateId :: Lens' FlowExecutionSummary (Maybe Text)

The ID of the flow.

flowExecutionSummary_createdAt :: Lens' FlowExecutionSummary (Maybe UTCTime)

The date and time when the flow execution summary was created.

flowExecutionSummary_flowExecutionId :: Lens' FlowExecutionSummary (Maybe Text)

The ID of the flow execution.

flowExecutionSummary_systemInstanceId :: Lens' FlowExecutionSummary (Maybe Text)

The ID of the system instance that contains the flow.

flowExecutionSummary_updatedAt :: Lens' FlowExecutionSummary (Maybe UTCTime)

The date and time when the flow execution summary was last updated.