Amazonka.IoTThingsGraph.Types.SystemTemplateDescription

Description

Documentation

data SystemTemplateDescription

An object that contains a system's definition document and summary information.

See: newSystemTemplateDescription smart constructor.

Constructors

SystemTemplateDescription'

Fields

summary :: Maybe SystemTemplateSummary

An object that contains summary information about a system.

definition :: Maybe DefinitionDocument

The definition document of a system.

validatedNamespaceVersion :: Maybe Integer

The namespace version against which the system was validated. Use this value in your system instance.

newSystemTemplateDescription :: SystemTemplateDescription

Create a value of SystemTemplateDescription with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:summary:SystemTemplateDescription', systemTemplateDescription_summary - An object that contains summary information about a system.

$sel:definition:SystemTemplateDescription', systemTemplateDescription_definition - The definition document of a system.

$sel:validatedNamespaceVersion:SystemTemplateDescription', systemTemplateDescription_validatedNamespaceVersion - The namespace version against which the system was validated. Use this value in your system instance.

systemTemplateDescription_summary :: Lens' SystemTemplateDescription (Maybe SystemTemplateSummary)

An object that contains summary information about a system.

systemTemplateDescription_definition :: Lens' SystemTemplateDescription (Maybe DefinitionDocument)

The definition document of a system.

systemTemplateDescription_validatedNamespaceVersion :: Lens' SystemTemplateDescription (Maybe Integer)

The namespace version against which the system was validated. Use this value in your system instance.