Amazonka.IoTThingsGraph.Types.SystemTemplateFilter

Description

Documentation

data SystemTemplateFilter

An object that filters a system search.

See: newSystemTemplateFilter smart constructor.

Constructors

SystemTemplateFilter'

Fields

name :: SystemTemplateFilterName

The name of the system search filter field.

value :: [Text]

An array of string values for the search filter field. Multiple values function as AND criteria in the search.

newSystemTemplateFilter

Arguments

:: SystemTemplateFilterName
$sel:name:SystemTemplateFilter'

-> SystemTemplateFilter

Create a value of SystemTemplateFilter with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:name:SystemTemplateFilter', systemTemplateFilter_name - The name of the system search filter field.

$sel:value:SystemTemplateFilter', systemTemplateFilter_value - An array of string values for the search filter field. Multiple values function as AND criteria in the search.

systemTemplateFilter_name :: Lens' SystemTemplateFilter SystemTemplateFilterName

The name of the system search filter field.

systemTemplateFilter_value :: Lens' SystemTemplateFilter [Text]

An array of string values for the search filter field. Multiple values function as AND criteria in the search.