Amazonka.IoTThingsGraph.UpdateFlowTemplate

Description

Updates the specified workflow. All deployed systems and system instances that use the workflow will see the changes in the flow when it is redeployed. If you don't want this behavior, copy the workflow (creating a new workflow with a different ID), and update the copy. The workflow can contain only entities in the specified namespace.

Creating a Request

data UpdateFlowTemplate

See: newUpdateFlowTemplate smart constructor.

Constructors

UpdateFlowTemplate'

Fields

compatibleNamespaceVersion :: Maybe Integer

The version of the user's namespace.

If no value is specified, the latest version is used by default. Use the GetFlowTemplateRevisions if you want to find earlier revisions of the flow to update.

id :: Text

The ID of the workflow to be updated.

The ID should be in the following format.

urn:tdm:REGION/ACCOUNT ID/default:workflow:WORKFLOWNAME

definition :: DefinitionDocument

The DefinitionDocument that contains the updated workflow definition.

newUpdateFlowTemplate

Arguments

:: Text
$sel:id:UpdateFlowTemplate'

-> DefinitionDocument
$sel:definition:UpdateFlowTemplate'

-> UpdateFlowTemplate

Create a value of UpdateFlowTemplate with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:compatibleNamespaceVersion:UpdateFlowTemplate', updateFlowTemplate_compatibleNamespaceVersion - The version of the user's namespace.

If no value is specified, the latest version is used by default. Use the GetFlowTemplateRevisions if you want to find earlier revisions of the flow to update.

$sel:id:UpdateFlowTemplate', updateFlowTemplate_id - The ID of the workflow to be updated.

The ID should be in the following format.

urn:tdm:REGION/ACCOUNT ID/default:workflow:WORKFLOWNAME

$sel:definition:UpdateFlowTemplate', updateFlowTemplate_definition - The DefinitionDocument that contains the updated workflow definition.

Request Lenses

updateFlowTemplate_compatibleNamespaceVersion :: Lens' UpdateFlowTemplate (Maybe Integer)

The version of the user's namespace.

If no value is specified, the latest version is used by default. Use the GetFlowTemplateRevisions if you want to find earlier revisions of the flow to update.

updateFlowTemplate_id :: Lens' UpdateFlowTemplate Text

The ID of the workflow to be updated.

The ID should be in the following format.

urn:tdm:REGION/ACCOUNT ID/default:workflow:WORKFLOWNAME

updateFlowTemplate_definition :: Lens' UpdateFlowTemplate DefinitionDocument

The DefinitionDocument that contains the updated workflow definition.

Destructuring the Response

data UpdateFlowTemplateResponse

See: newUpdateFlowTemplateResponse smart constructor.

Constructors

UpdateFlowTemplateResponse'

Fields

summary :: Maybe FlowTemplateSummary

An object containing summary information about the updated workflow.

httpStatus :: Int

The response's http status code.

newUpdateFlowTemplateResponse

Arguments

:: Int
$sel:httpStatus:UpdateFlowTemplateResponse'

-> UpdateFlowTemplateResponse

Create a value of UpdateFlowTemplateResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:summary:UpdateFlowTemplateResponse', updateFlowTemplateResponse_summary - An object containing summary information about the updated workflow.

$sel:httpStatus:UpdateFlowTemplateResponse', updateFlowTemplateResponse_httpStatus - The response's http status code.

Response Lenses

updateFlowTemplateResponse_summary :: Lens' UpdateFlowTemplateResponse (Maybe FlowTemplateSummary)

An object containing summary information about the updated workflow.

updateFlowTemplateResponse_httpStatus :: Lens' UpdateFlowTemplateResponse Int

The response's http status code.