Amazonka.IoTWireless.GetPartnerAccount

Description

Gets information about a partner account. If PartnerAccountId and PartnerType are null, returns all partner accounts.

Creating a Request

data GetPartnerAccount

See: newGetPartnerAccount smart constructor.

Constructors

GetPartnerAccount'

Fields

partnerAccountId :: Text

The partner account ID to disassociate from the AWS account.

partnerType :: PartnerType

The partner type.

newGetPartnerAccount

Arguments

:: Text
$sel:partnerAccountId:GetPartnerAccount'

-> PartnerType
$sel:partnerType:GetPartnerAccount'

-> GetPartnerAccount

Create a value of GetPartnerAccount with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:partnerAccountId:GetPartnerAccount', getPartnerAccount_partnerAccountId - The partner account ID to disassociate from the AWS account.

$sel:partnerType:GetPartnerAccount', getPartnerAccount_partnerType - The partner type.

Request Lenses

getPartnerAccount_partnerAccountId :: Lens' GetPartnerAccount Text

The partner account ID to disassociate from the AWS account.

getPartnerAccount_partnerType :: Lens' GetPartnerAccount PartnerType

The partner type.

Destructuring the Response

data GetPartnerAccountResponse

See: newGetPartnerAccountResponse smart constructor.

Constructors

GetPartnerAccountResponse'

Fields

sidewalk :: Maybe SidewalkAccountInfoWithFingerprint

The Sidewalk account credentials.

accountLinked :: Maybe Bool

Whether the partner account is linked to the AWS account.

httpStatus :: Int

The response's http status code.

newGetPartnerAccountResponse

Arguments

:: Int
$sel:httpStatus:GetPartnerAccountResponse'

-> GetPartnerAccountResponse

Create a value of GetPartnerAccountResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:sidewalk:GetPartnerAccountResponse', getPartnerAccountResponse_sidewalk - The Sidewalk account credentials.

$sel:accountLinked:GetPartnerAccountResponse', getPartnerAccountResponse_accountLinked - Whether the partner account is linked to the AWS account.

$sel:httpStatus:GetPartnerAccountResponse', getPartnerAccountResponse_httpStatus - The response's http status code.

Response Lenses

getPartnerAccountResponse_sidewalk :: Lens' GetPartnerAccountResponse (Maybe SidewalkAccountInfoWithFingerprint)

The Sidewalk account credentials.

getPartnerAccountResponse_accountLinked :: Lens' GetPartnerAccountResponse (Maybe Bool)

Whether the partner account is linked to the AWS account.

getPartnerAccountResponse_httpStatus :: Lens' GetPartnerAccountResponse Int

The response's http status code.