Amazonka.IoTWireless.GetServiceEndpoint

Description

Gets the account-specific endpoint for Configuration and Update Server (CUPS) protocol or LoRaWAN Network Server (LNS) connections.

Creating a Request

data GetServiceEndpoint

See: newGetServiceEndpoint smart constructor.

Constructors

GetServiceEndpoint'

Fields

serviceType :: Maybe WirelessGatewayServiceType

The service type for which to get endpoint information about. Can be CUPS for the Configuration and Update Server endpoint, or LNS for the LoRaWAN Network Server endpoint.

newGetServiceEndpoint :: GetServiceEndpoint

Create a value of GetServiceEndpoint with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:serviceType:GetServiceEndpoint', getServiceEndpoint_serviceType - The service type for which to get endpoint information about. Can be CUPS for the Configuration and Update Server endpoint, or LNS for the LoRaWAN Network Server endpoint.

Request Lenses

getServiceEndpoint_serviceType :: Lens' GetServiceEndpoint (Maybe WirelessGatewayServiceType)

The service type for which to get endpoint information about. Can be CUPS for the Configuration and Update Server endpoint, or LNS for the LoRaWAN Network Server endpoint.

Destructuring the Response

data GetServiceEndpointResponse

See: newGetServiceEndpointResponse smart constructor.

Constructors

GetServiceEndpointResponse'

Fields

serviceEndpoint :: Maybe Text

The service endpoint value.

serverTrust :: Maybe Text

The Root CA of the server trust certificate.

serviceType :: Maybe WirelessGatewayServiceType

The endpoint's service type.

httpStatus :: Int

The response's http status code.

newGetServiceEndpointResponse

Arguments

:: Int
$sel:httpStatus:GetServiceEndpointResponse'

-> GetServiceEndpointResponse

Create a value of GetServiceEndpointResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:serviceEndpoint:GetServiceEndpointResponse', getServiceEndpointResponse_serviceEndpoint - The service endpoint value.

$sel:serverTrust:GetServiceEndpointResponse', getServiceEndpointResponse_serverTrust - The Root CA of the server trust certificate.

$sel:serviceType:GetServiceEndpoint', getServiceEndpointResponse_serviceType - The endpoint's service type.

$sel:httpStatus:GetServiceEndpointResponse', getServiceEndpointResponse_httpStatus - The response's http status code.

Response Lenses

getServiceEndpointResponse_serviceEndpoint :: Lens' GetServiceEndpointResponse (Maybe Text)

The service endpoint value.

getServiceEndpointResponse_serverTrust :: Lens' GetServiceEndpointResponse (Maybe Text)

The Root CA of the server trust certificate.

getServiceEndpointResponse_serviceType :: Lens' GetServiceEndpointResponse (Maybe WirelessGatewayServiceType)

The endpoint's service type.

getServiceEndpointResponse_httpStatus :: Lens' GetServiceEndpointResponse Int

The response's http status code.