Amazonka.IoTWireless.GetWirelessDevice

Description

Gets information about a wireless device.

Creating a Request

data GetWirelessDevice

See: newGetWirelessDevice smart constructor.

Constructors

GetWirelessDevice'

Fields

identifier :: Text

The identifier of the wireless device to get.

identifierType :: WirelessDeviceIdType

The type of identifier used in identifier.

newGetWirelessDevice

Arguments

:: Text
$sel:identifier:GetWirelessDevice'

-> WirelessDeviceIdType
$sel:identifierType:GetWirelessDevice'

-> GetWirelessDevice

Create a value of GetWirelessDevice with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:identifier:GetWirelessDevice', getWirelessDevice_identifier - The identifier of the wireless device to get.

$sel:identifierType:GetWirelessDevice', getWirelessDevice_identifierType - The type of identifier used in identifier.

Request Lenses

getWirelessDevice_identifier :: Lens' GetWirelessDevice Text

The identifier of the wireless device to get.

getWirelessDevice_identifierType :: Lens' GetWirelessDevice WirelessDeviceIdType

The type of identifier used in identifier.

Destructuring the Response

data GetWirelessDeviceResponse

See: newGetWirelessDeviceResponse smart constructor.

Constructors

GetWirelessDeviceResponse'

Fields

sidewalk :: Maybe SidewalkDevice

Sidewalk device object.

arn :: Maybe Text

The Amazon Resource Name of the resource.

loRaWAN :: Maybe LoRaWANDevice

Information about the wireless device.

thingArn :: Maybe Text

The ARN of the thing associated with the wireless device.

name :: Maybe Text

The name of the resource.

id :: Maybe Text

The ID of the wireless device.

type' :: Maybe WirelessDeviceType

The wireless device type.

destinationName :: Maybe Text

The name of the destination to which the device is assigned.

thingName :: Maybe Text

The name of the thing associated with the wireless device. The value is empty if a thing isn't associated with the device.

description :: Maybe Text

The description of the resource.

httpStatus :: Int

The response's http status code.

newGetWirelessDeviceResponse

Arguments

:: Int
$sel:httpStatus:GetWirelessDeviceResponse'

-> GetWirelessDeviceResponse

Create a value of GetWirelessDeviceResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:sidewalk:GetWirelessDeviceResponse', getWirelessDeviceResponse_sidewalk - Sidewalk device object.

$sel:arn:GetWirelessDeviceResponse', getWirelessDeviceResponse_arn - The Amazon Resource Name of the resource.

$sel:loRaWAN:GetWirelessDeviceResponse', getWirelessDeviceResponse_loRaWAN - Information about the wireless device.

$sel:thingArn:GetWirelessDeviceResponse', getWirelessDeviceResponse_thingArn - The ARN of the thing associated with the wireless device.

$sel:name:GetWirelessDeviceResponse', getWirelessDeviceResponse_name - The name of the resource.

$sel:id:GetWirelessDeviceResponse', getWirelessDeviceResponse_id - The ID of the wireless device.

$sel:type':GetWirelessDeviceResponse', getWirelessDeviceResponse_type - The wireless device type.

$sel:destinationName:GetWirelessDeviceResponse', getWirelessDeviceResponse_destinationName - The name of the destination to which the device is assigned.

$sel:thingName:GetWirelessDeviceResponse', getWirelessDeviceResponse_thingName - The name of the thing associated with the wireless device. The value is empty if a thing isn't associated with the device.

$sel:description:GetWirelessDeviceResponse', getWirelessDeviceResponse_description - The description of the resource.

$sel:httpStatus:GetWirelessDeviceResponse', getWirelessDeviceResponse_httpStatus - The response's http status code.

Response Lenses

getWirelessDeviceResponse_sidewalk :: Lens' GetWirelessDeviceResponse (Maybe SidewalkDevice)

Sidewalk device object.

getWirelessDeviceResponse_arn :: Lens' GetWirelessDeviceResponse (Maybe Text)

The Amazon Resource Name of the resource.

getWirelessDeviceResponse_loRaWAN :: Lens' GetWirelessDeviceResponse (Maybe LoRaWANDevice)

Information about the wireless device.

getWirelessDeviceResponse_thingArn :: Lens' GetWirelessDeviceResponse (Maybe Text)

The ARN of the thing associated with the wireless device.

getWirelessDeviceResponse_name :: Lens' GetWirelessDeviceResponse (Maybe Text)

The name of the resource.

getWirelessDeviceResponse_id :: Lens' GetWirelessDeviceResponse (Maybe Text)

The ID of the wireless device.

getWirelessDeviceResponse_type :: Lens' GetWirelessDeviceResponse (Maybe WirelessDeviceType)

The wireless device type.

getWirelessDeviceResponse_destinationName :: Lens' GetWirelessDeviceResponse (Maybe Text)

The name of the destination to which the device is assigned.

getWirelessDeviceResponse_thingName :: Lens' GetWirelessDeviceResponse (Maybe Text)

The name of the thing associated with the wireless device. The value is empty if a thing isn't associated with the device.

getWirelessDeviceResponse_description :: Lens' GetWirelessDeviceResponse (Maybe Text)

The description of the resource.

getWirelessDeviceResponse_httpStatus :: Lens' GetWirelessDeviceResponse Int

The response's http status code.