Amazonka.IoTWireless.GetWirelessGateway

Description

Gets information about a wireless gateway.

Creating a Request

data GetWirelessGateway

See: newGetWirelessGateway smart constructor.

Constructors

GetWirelessGateway'

Fields

identifier :: Text

The identifier of the wireless gateway to get.

identifierType :: WirelessGatewayIdType

The type of identifier used in identifier.

newGetWirelessGateway

Arguments

:: Text
$sel:identifier:GetWirelessGateway'

-> WirelessGatewayIdType
$sel:identifierType:GetWirelessGateway'

-> GetWirelessGateway

Create a value of GetWirelessGateway with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:identifier:GetWirelessGateway', getWirelessGateway_identifier - The identifier of the wireless gateway to get.

$sel:identifierType:GetWirelessGateway', getWirelessGateway_identifierType - The type of identifier used in identifier.

Request Lenses

getWirelessGateway_identifier :: Lens' GetWirelessGateway Text

The identifier of the wireless gateway to get.

getWirelessGateway_identifierType :: Lens' GetWirelessGateway WirelessGatewayIdType

The type of identifier used in identifier.

Destructuring the Response

data GetWirelessGatewayResponse

See: newGetWirelessGatewayResponse smart constructor.

Constructors

GetWirelessGatewayResponse'

Fields

arn :: Maybe Text

The Amazon Resource Name of the resource.

loRaWAN :: Maybe LoRaWANGateway

Information about the wireless gateway.

thingArn :: Maybe Text

The ARN of the thing associated with the wireless gateway.

name :: Maybe Text

The name of the resource.

id :: Maybe Text

The ID of the wireless gateway.

thingName :: Maybe Text

The name of the thing associated with the wireless gateway. The value is empty if a thing isn't associated with the gateway.

description :: Maybe Text

The description of the resource.

httpStatus :: Int

The response's http status code.

newGetWirelessGatewayResponse

Arguments

:: Int
$sel:httpStatus:GetWirelessGatewayResponse'

-> GetWirelessGatewayResponse

Create a value of GetWirelessGatewayResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:arn:GetWirelessGatewayResponse', getWirelessGatewayResponse_arn - The Amazon Resource Name of the resource.

$sel:loRaWAN:GetWirelessGatewayResponse', getWirelessGatewayResponse_loRaWAN - Information about the wireless gateway.

$sel:thingArn:GetWirelessGatewayResponse', getWirelessGatewayResponse_thingArn - The ARN of the thing associated with the wireless gateway.

$sel:name:GetWirelessGatewayResponse', getWirelessGatewayResponse_name - The name of the resource.

$sel:id:GetWirelessGatewayResponse', getWirelessGatewayResponse_id - The ID of the wireless gateway.

$sel:thingName:GetWirelessGatewayResponse', getWirelessGatewayResponse_thingName - The name of the thing associated with the wireless gateway. The value is empty if a thing isn't associated with the gateway.

$sel:description:GetWirelessGatewayResponse', getWirelessGatewayResponse_description - The description of the resource.

$sel:httpStatus:GetWirelessGatewayResponse', getWirelessGatewayResponse_httpStatus - The response's http status code.

Response Lenses

getWirelessGatewayResponse_arn :: Lens' GetWirelessGatewayResponse (Maybe Text)

The Amazon Resource Name of the resource.

getWirelessGatewayResponse_loRaWAN :: Lens' GetWirelessGatewayResponse (Maybe LoRaWANGateway)

Information about the wireless gateway.

getWirelessGatewayResponse_thingArn :: Lens' GetWirelessGatewayResponse (Maybe Text)

The ARN of the thing associated with the wireless gateway.

getWirelessGatewayResponse_name :: Lens' GetWirelessGatewayResponse (Maybe Text)

The name of the resource.

getWirelessGatewayResponse_id :: Lens' GetWirelessGatewayResponse (Maybe Text)

The ID of the wireless gateway.

getWirelessGatewayResponse_thingName :: Lens' GetWirelessGatewayResponse (Maybe Text)

The name of the thing associated with the wireless gateway. The value is empty if a thing isn't associated with the gateway.

getWirelessGatewayResponse_description :: Lens' GetWirelessGatewayResponse (Maybe Text)

The description of the resource.

getWirelessGatewayResponse_httpStatus :: Lens' GetWirelessGatewayResponse Int

The response's http status code.