Amazonka.IoTWireless.GetWirelessGatewayTask

Description

Gets information about a wireless gateway task.

Creating a Request

data GetWirelessGatewayTask

See: newGetWirelessGatewayTask smart constructor.

Constructors

GetWirelessGatewayTask'

Fields

id :: Text

The ID of the resource to get.

newGetWirelessGatewayTask

Arguments

:: Text
$sel:id:GetWirelessGatewayTask'

-> GetWirelessGatewayTask

Create a value of GetWirelessGatewayTask with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:id:GetWirelessGatewayTask', getWirelessGatewayTask_id - The ID of the resource to get.

Request Lenses

getWirelessGatewayTask_id :: Lens' GetWirelessGatewayTask Text

The ID of the resource to get.

Destructuring the Response

data GetWirelessGatewayTaskResponse

See: newGetWirelessGatewayTaskResponse smart constructor.

Constructors

GetWirelessGatewayTaskResponse'

Fields

status :: Maybe WirelessGatewayTaskStatus

The status of the request.

wirelessGatewayTaskDefinitionId :: Maybe Text

The ID of the WirelessGatewayTask.

taskCreatedAt :: Maybe Text

The date and time when the task was created.

lastUplinkReceivedAt :: Maybe Text

The date and time when the most recent uplink was received.

wirelessGatewayId :: Maybe Text

The ID of the wireless gateway.

httpStatus :: Int

The response's http status code.

newGetWirelessGatewayTaskResponse

Arguments

:: Int
$sel:httpStatus:GetWirelessGatewayTaskResponse'

-> GetWirelessGatewayTaskResponse

Create a value of GetWirelessGatewayTaskResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:status:GetWirelessGatewayTaskResponse', getWirelessGatewayTaskResponse_status - The status of the request.

$sel:wirelessGatewayTaskDefinitionId:GetWirelessGatewayTaskResponse', getWirelessGatewayTaskResponse_wirelessGatewayTaskDefinitionId - The ID of the WirelessGatewayTask.

$sel:taskCreatedAt:GetWirelessGatewayTaskResponse', getWirelessGatewayTaskResponse_taskCreatedAt - The date and time when the task was created.

$sel:lastUplinkReceivedAt:GetWirelessGatewayTaskResponse', getWirelessGatewayTaskResponse_lastUplinkReceivedAt - The date and time when the most recent uplink was received.

$sel:wirelessGatewayId:GetWirelessGatewayTaskResponse', getWirelessGatewayTaskResponse_wirelessGatewayId - The ID of the wireless gateway.

$sel:httpStatus:GetWirelessGatewayTaskResponse', getWirelessGatewayTaskResponse_httpStatus - The response's http status code.

Response Lenses

getWirelessGatewayTaskResponse_status :: Lens' GetWirelessGatewayTaskResponse (Maybe WirelessGatewayTaskStatus)

The status of the request.

getWirelessGatewayTaskResponse_wirelessGatewayTaskDefinitionId :: Lens' GetWirelessGatewayTaskResponse (Maybe Text)

The ID of the WirelessGatewayTask.

getWirelessGatewayTaskResponse_taskCreatedAt :: Lens' GetWirelessGatewayTaskResponse (Maybe Text)

The date and time when the task was created.

getWirelessGatewayTaskResponse_lastUplinkReceivedAt :: Lens' GetWirelessGatewayTaskResponse (Maybe Text)

The date and time when the most recent uplink was received.

getWirelessGatewayTaskResponse_wirelessGatewayId :: Lens' GetWirelessGatewayTaskResponse (Maybe Text)

The ID of the wireless gateway.

getWirelessGatewayTaskResponse_httpStatus :: Lens' GetWirelessGatewayTaskResponse Int

The response's http status code.