Amazonka.IoTWireless.ListWirelessGateways

Description

Lists the wireless gateways registered to your AWS account.

Creating a Request

data ListWirelessGateways

See: newListWirelessGateways smart constructor.

Constructors

ListWirelessGateways'

Fields

nextToken :: Maybe Text

To retrieve the next set of results, the nextToken value from a previous response; otherwise null to receive the first set of results.

maxResults :: Maybe Natural

The maximum number of results to return in this operation.

newListWirelessGateways :: ListWirelessGateways

Create a value of ListWirelessGateways with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:nextToken:ListWirelessGateways', listWirelessGateways_nextToken - To retrieve the next set of results, the nextToken value from a previous response; otherwise null to receive the first set of results.

$sel:maxResults:ListWirelessGateways', listWirelessGateways_maxResults - The maximum number of results to return in this operation.

Request Lenses

listWirelessGateways_nextToken :: Lens' ListWirelessGateways (Maybe Text)

To retrieve the next set of results, the nextToken value from a previous response; otherwise null to receive the first set of results.

listWirelessGateways_maxResults :: Lens' ListWirelessGateways (Maybe Natural)

The maximum number of results to return in this operation.

Destructuring the Response

data ListWirelessGatewaysResponse

See: newListWirelessGatewaysResponse smart constructor.

Constructors

ListWirelessGatewaysResponse'

Fields

wirelessGatewayList :: Maybe [WirelessGatewayStatistics]

The ID of the wireless gateway.

nextToken :: Maybe Text

The token to use to get the next set of results, or null if there are no additional results.

httpStatus :: Int

The response's http status code.

newListWirelessGatewaysResponse

Arguments

:: Int
$sel:httpStatus:ListWirelessGatewaysResponse'

-> ListWirelessGatewaysResponse

Create a value of ListWirelessGatewaysResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:wirelessGatewayList:ListWirelessGatewaysResponse', listWirelessGatewaysResponse_wirelessGatewayList - The ID of the wireless gateway.

$sel:nextToken:ListWirelessGateways', listWirelessGatewaysResponse_nextToken - The token to use to get the next set of results, or null if there are no additional results.

$sel:httpStatus:ListWirelessGatewaysResponse', listWirelessGatewaysResponse_httpStatus - The response's http status code.

Response Lenses

listWirelessGatewaysResponse_wirelessGatewayList :: Lens' ListWirelessGatewaysResponse (Maybe [WirelessGatewayStatistics])

The ID of the wireless gateway.

listWirelessGatewaysResponse_nextToken :: Lens' ListWirelessGatewaysResponse (Maybe Text)

The token to use to get the next set of results, or null if there are no additional results.

listWirelessGatewaysResponse_httpStatus :: Lens' ListWirelessGatewaysResponse Int

The response's http status code.