Amazonka.IoTWireless.SendDataToWirelessDevice

Description

Sends a decrypted application data frame to a device.

Creating a Request

data SendDataToWirelessDevice

See: newSendDataToWirelessDevice smart constructor.

Constructors

SendDataToWirelessDevice'

Fields

wirelessMetadata :: Maybe WirelessMetadata

Metadata about the message request.

id :: Text

The ID of the wireless device to receive the data.

transmitMode :: Natural

The transmit mode to use to send data to the wireless device. Can be: 0 for UM (unacknowledge mode) or 1 for AM (acknowledge mode).

payloadData :: Text

The binary to be sent to the end device, encoded in base64.

newSendDataToWirelessDevice

Arguments

:: Text
$sel:id:SendDataToWirelessDevice'

-> Natural
$sel:transmitMode:SendDataToWirelessDevice'

-> Text
$sel:payloadData:SendDataToWirelessDevice'

-> SendDataToWirelessDevice

Create a value of SendDataToWirelessDevice with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:wirelessMetadata:SendDataToWirelessDevice', sendDataToWirelessDevice_wirelessMetadata - Metadata about the message request.

$sel:id:SendDataToWirelessDevice', sendDataToWirelessDevice_id - The ID of the wireless device to receive the data.

$sel:transmitMode:SendDataToWirelessDevice', sendDataToWirelessDevice_transmitMode - The transmit mode to use to send data to the wireless device. Can be: 0 for UM (unacknowledge mode) or 1 for AM (acknowledge mode).

$sel:payloadData:SendDataToWirelessDevice', sendDataToWirelessDevice_payloadData - The binary to be sent to the end device, encoded in base64.

Request Lenses

sendDataToWirelessDevice_wirelessMetadata :: Lens' SendDataToWirelessDevice (Maybe WirelessMetadata)

Metadata about the message request.

sendDataToWirelessDevice_id :: Lens' SendDataToWirelessDevice Text

The ID of the wireless device to receive the data.

sendDataToWirelessDevice_transmitMode :: Lens' SendDataToWirelessDevice Natural

The transmit mode to use to send data to the wireless device. Can be: 0 for UM (unacknowledge mode) or 1 for AM (acknowledge mode).

sendDataToWirelessDevice_payloadData :: Lens' SendDataToWirelessDevice Text

The binary to be sent to the end device, encoded in base64.

Destructuring the Response

data SendDataToWirelessDeviceResponse

See: newSendDataToWirelessDeviceResponse smart constructor.

Constructors

SendDataToWirelessDeviceResponse'

Fields

messageId :: Maybe Text

The ID of the message sent to the wireless device.

httpStatus :: Int

The response's http status code.

newSendDataToWirelessDeviceResponse

Arguments

:: Int
$sel:httpStatus:SendDataToWirelessDeviceResponse'

-> SendDataToWirelessDeviceResponse

Create a value of SendDataToWirelessDeviceResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:messageId:SendDataToWirelessDeviceResponse', sendDataToWirelessDeviceResponse_messageId - The ID of the message sent to the wireless device.

$sel:httpStatus:SendDataToWirelessDeviceResponse', sendDataToWirelessDeviceResponse_httpStatus - The response's http status code.

Response Lenses

sendDataToWirelessDeviceResponse_messageId :: Lens' SendDataToWirelessDeviceResponse (Maybe Text)

The ID of the message sent to the wireless device.

sendDataToWirelessDeviceResponse_httpStatus :: Lens' SendDataToWirelessDeviceResponse Int

The response's http status code.