Amazonka.IoTWireless.Types.LoRaWANDevice

Description

Documentation

data LoRaWANDevice

LoRaWAN object for create functions.

See: newLoRaWANDevice smart constructor.

Constructors

LoRaWANDevice'

Fields

abpV1_0_x :: Maybe AbpV1_0_x

LoRaWAN object for create APIs

otaaV1_1 :: Maybe OtaaV1_1

OTAA device object for v1.1 for create APIs

serviceProfileId :: Maybe Text

The ID of the service profile.

deviceProfileId :: Maybe Text

The ID of the device profile for the new wireless device.

otaaV1_0_x :: Maybe OtaaV1_0_x

OTAA device object for create APIs for v1.0.x

abpV1_1 :: Maybe AbpV1_1

ABP device object for create APIs for v1.1

devEui :: Maybe Text

The DevEUI value.

newLoRaWANDevice :: LoRaWANDevice

Create a value of LoRaWANDevice with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:abpV1_0_x:LoRaWANDevice', loRaWANDevice_abpV1_0_x - LoRaWAN object for create APIs

$sel:otaaV1_1:LoRaWANDevice', loRaWANDevice_otaaV1_1 - OTAA device object for v1.1 for create APIs

$sel:serviceProfileId:LoRaWANDevice', loRaWANDevice_serviceProfileId - The ID of the service profile.

$sel:deviceProfileId:LoRaWANDevice', loRaWANDevice_deviceProfileId - The ID of the device profile for the new wireless device.

$sel:otaaV1_0_x:LoRaWANDevice', loRaWANDevice_otaaV1_0_x - OTAA device object for create APIs for v1.0.x

$sel:abpV1_1:LoRaWANDevice', loRaWANDevice_abpV1_1 - ABP device object for create APIs for v1.1

$sel:devEui:LoRaWANDevice', loRaWANDevice_devEui - The DevEUI value.

loRaWANDevice_abpV1_0_x :: Lens' LoRaWANDevice (Maybe AbpV1_0_x)

LoRaWAN object for create APIs

loRaWANDevice_otaaV1_1 :: Lens' LoRaWANDevice (Maybe OtaaV1_1)

OTAA device object for v1.1 for create APIs

loRaWANDevice_serviceProfileId :: Lens' LoRaWANDevice (Maybe Text)

The ID of the service profile.

loRaWANDevice_deviceProfileId :: Lens' LoRaWANDevice (Maybe Text)

The ID of the device profile for the new wireless device.

loRaWANDevice_otaaV1_0_x :: Lens' LoRaWANDevice (Maybe OtaaV1_0_x)

OTAA device object for create APIs for v1.0.x

loRaWANDevice_abpV1_1 :: Lens' LoRaWANDevice (Maybe AbpV1_1)

ABP device object for create APIs for v1.1

loRaWANDevice_devEui :: Lens' LoRaWANDevice (Maybe Text)

The DevEUI value.