Amazonka.IoTWireless.Types.WirelessDeviceEventLogOption

Description

Documentation

data WirelessDeviceEventLogOption

The log options for a wireless device event and can be used to set log levels for a specific wireless device event.

For a LoRaWAN device, possible events for a log messsage are: Join, Rejoin, Downlink_Data, and Uplink_Data. For a Sidewalk device, possible events for a log message are Registration, Downlink_Data, and Uplink_Data.

See: newWirelessDeviceEventLogOption smart constructor.

Constructors

WirelessDeviceEventLogOption'

Fields

event :: WirelessDeviceEvent

logLevel :: LogLevel

newWirelessDeviceEventLogOption

Arguments

:: WirelessDeviceEvent
$sel:event:WirelessDeviceEventLogOption'

-> LogLevel
$sel:logLevel:WirelessDeviceEventLogOption'

-> WirelessDeviceEventLogOption

Create a value of WirelessDeviceEventLogOption with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:event:WirelessDeviceEventLogOption', wirelessDeviceEventLogOption_event - Undocumented member.

$sel:logLevel:WirelessDeviceEventLogOption', wirelessDeviceEventLogOption_logLevel - Undocumented member.

wirelessDeviceEventLogOption_event :: Lens' WirelessDeviceEventLogOption WirelessDeviceEvent

Undocumented member.

wirelessDeviceEventLogOption_logLevel :: Lens' WirelessDeviceEventLogOption LogLevel

Undocumented member.