Amazonka.IoTWireless.Types.WirelessDeviceLogOption

Description

Documentation

data WirelessDeviceLogOption

The log options for wireless devices and can be used to set log levels for a specific type of wireless device.

See: newWirelessDeviceLogOption smart constructor.

Constructors

WirelessDeviceLogOption'

Fields

events :: Maybe [WirelessDeviceEventLogOption]

type' :: WirelessDeviceType

The wireless device type.

logLevel :: LogLevel

newWirelessDeviceLogOption

Arguments

:: WirelessDeviceType
$sel:type':WirelessDeviceLogOption'

-> LogLevel
$sel:logLevel:WirelessDeviceLogOption'

-> WirelessDeviceLogOption

Create a value of WirelessDeviceLogOption with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:events:WirelessDeviceLogOption', wirelessDeviceLogOption_events - Undocumented member.

$sel:type':WirelessDeviceLogOption', wirelessDeviceLogOption_type - The wireless device type.

$sel:logLevel:WirelessDeviceLogOption', wirelessDeviceLogOption_logLevel - Undocumented member.

wirelessDeviceLogOption_events :: Lens' WirelessDeviceLogOption (Maybe [WirelessDeviceEventLogOption])

Undocumented member.

wirelessDeviceLogOption_type :: Lens' WirelessDeviceLogOption WirelessDeviceType

The wireless device type.

wirelessDeviceLogOption_logLevel :: Lens' WirelessDeviceLogOption LogLevel

Undocumented member.