Amazonka.IoTWireless.Types.WirelessGatewayLogOption

Description

Documentation

data WirelessGatewayLogOption

The log options for wireless gateways and can be used to set log levels for a specific type of wireless gateway.

See: newWirelessGatewayLogOption smart constructor.

Constructors

WirelessGatewayLogOption'

Fields

events :: Maybe [WirelessGatewayEventLogOption]

type' :: WirelessGatewayType

logLevel :: LogLevel

newWirelessGatewayLogOption

Arguments

:: WirelessGatewayType
$sel:type':WirelessGatewayLogOption'

-> LogLevel
$sel:logLevel:WirelessGatewayLogOption'

-> WirelessGatewayLogOption

Create a value of WirelessGatewayLogOption with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:events:WirelessGatewayLogOption', wirelessGatewayLogOption_events - Undocumented member.

$sel:type':WirelessGatewayLogOption', wirelessGatewayLogOption_type - Undocumented member.

$sel:logLevel:WirelessGatewayLogOption', wirelessGatewayLogOption_logLevel - Undocumented member.

wirelessGatewayLogOption_events :: Lens' WirelessGatewayLogOption (Maybe [WirelessGatewayEventLogOption])

Undocumented member.

wirelessGatewayLogOption_type :: Lens' WirelessGatewayLogOption WirelessGatewayType

Undocumented member.

wirelessGatewayLogOption_logLevel :: Lens' WirelessGatewayLogOption LogLevel

Undocumented member.