Amazonka.IoTWireless.Types.WirelessGatewayStatistics

Description

Documentation

data WirelessGatewayStatistics

Information about a wireless gateway's operation.

See: newWirelessGatewayStatistics smart constructor.

Constructors

WirelessGatewayStatistics'

Fields

arn :: Maybe Text

The Amazon Resource Name of the resource.

loRaWAN :: Maybe LoRaWANGateway

LoRaWAN gateway info.

name :: Maybe Text

The name of the resource.

id :: Maybe Text

The ID of the wireless gateway reporting the data.

lastUplinkReceivedAt :: Maybe Text

The date and time when the most recent uplink was received.

description :: Maybe Text

The description of the resource.

newWirelessGatewayStatistics :: WirelessGatewayStatistics

Create a value of WirelessGatewayStatistics with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:arn:WirelessGatewayStatistics', wirelessGatewayStatistics_arn - The Amazon Resource Name of the resource.

$sel:loRaWAN:WirelessGatewayStatistics', wirelessGatewayStatistics_loRaWAN - LoRaWAN gateway info.

$sel:name:WirelessGatewayStatistics', wirelessGatewayStatistics_name - The name of the resource.

$sel:id:WirelessGatewayStatistics', wirelessGatewayStatistics_id - The ID of the wireless gateway reporting the data.

$sel:lastUplinkReceivedAt:WirelessGatewayStatistics', wirelessGatewayStatistics_lastUplinkReceivedAt - The date and time when the most recent uplink was received.

$sel:description:WirelessGatewayStatistics', wirelessGatewayStatistics_description - The description of the resource.

wirelessGatewayStatistics_arn :: Lens' WirelessGatewayStatistics (Maybe Text)

The Amazon Resource Name of the resource.

wirelessGatewayStatistics_loRaWAN :: Lens' WirelessGatewayStatistics (Maybe LoRaWANGateway)

LoRaWAN gateway info.

wirelessGatewayStatistics_name :: Lens' WirelessGatewayStatistics (Maybe Text)

The name of the resource.

wirelessGatewayStatistics_id :: Lens' WirelessGatewayStatistics (Maybe Text)

The ID of the wireless gateway reporting the data.

wirelessGatewayStatistics_lastUplinkReceivedAt :: Lens' WirelessGatewayStatistics (Maybe Text)

The date and time when the most recent uplink was received.

wirelessGatewayStatistics_description :: Lens' WirelessGatewayStatistics (Maybe Text)

The description of the resource.