Amazonka.IoTWireless.UpdateLogLevelsByResourceTypes

Description

Set default log level, or log levels by resource types. This can be for wireless device log options or wireless gateways log options and is used to control the log messages that'll be displayed in CloudWatch.

Creating a Request

data UpdateLogLevelsByResourceTypes

See: newUpdateLogLevelsByResourceTypes smart constructor.

Constructors

UpdateLogLevelsByResourceTypes'

Fields

defaultLogLevel :: Maybe LogLevel

wirelessGatewayLogOptions :: Maybe [WirelessGatewayLogOption]

wirelessDeviceLogOptions :: Maybe [WirelessDeviceLogOption]

newUpdateLogLevelsByResourceTypes :: UpdateLogLevelsByResourceTypes

Create a value of UpdateLogLevelsByResourceTypes with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:defaultLogLevel:UpdateLogLevelsByResourceTypes', updateLogLevelsByResourceTypes_defaultLogLevel - Undocumented member.

$sel:wirelessGatewayLogOptions:UpdateLogLevelsByResourceTypes', updateLogLevelsByResourceTypes_wirelessGatewayLogOptions - Undocumented member.

$sel:wirelessDeviceLogOptions:UpdateLogLevelsByResourceTypes', updateLogLevelsByResourceTypes_wirelessDeviceLogOptions - Undocumented member.

Request Lenses

updateLogLevelsByResourceTypes_defaultLogLevel :: Lens' UpdateLogLevelsByResourceTypes (Maybe LogLevel)

Undocumented member.

updateLogLevelsByResourceTypes_wirelessGatewayLogOptions :: Lens' UpdateLogLevelsByResourceTypes (Maybe [WirelessGatewayLogOption])

Undocumented member.

updateLogLevelsByResourceTypes_wirelessDeviceLogOptions :: Lens' UpdateLogLevelsByResourceTypes (Maybe [WirelessDeviceLogOption])

Undocumented member.

Destructuring the Response

data UpdateLogLevelsByResourceTypesResponse

See: newUpdateLogLevelsByResourceTypesResponse smart constructor.

Constructors

UpdateLogLevelsByResourceTypesResponse'

Fields

httpStatus :: Int

The response's http status code.

newUpdateLogLevelsByResourceTypesResponse

Arguments

:: Int
$sel:httpStatus:UpdateLogLevelsByResourceTypesResponse'

-> UpdateLogLevelsByResourceTypesResponse

Create a value of UpdateLogLevelsByResourceTypesResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:httpStatus:UpdateLogLevelsByResourceTypesResponse', updateLogLevelsByResourceTypesResponse_httpStatus - The response's http status code.

Response Lenses

updateLogLevelsByResourceTypesResponse_httpStatus :: Lens' UpdateLogLevelsByResourceTypesResponse Int

The response's http status code.