Amazonka.IoTWireless.UpdateWirelessDevice

Description

Updates properties of a wireless device.

Creating a Request

data UpdateWirelessDevice

See: newUpdateWirelessDevice smart constructor.

Constructors

UpdateWirelessDevice'

Fields

loRaWAN :: Maybe LoRaWANUpdateDevice

The updated wireless device's configuration.

name :: Maybe Text

The new name of the resource.

destinationName :: Maybe Text

The name of the new destination for the device.

description :: Maybe Text

A new description of the resource.

id :: Text

The ID of the resource to update.

newUpdateWirelessDevice

Arguments

:: Text
$sel:id:UpdateWirelessDevice'

-> UpdateWirelessDevice

Create a value of UpdateWirelessDevice with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:loRaWAN:UpdateWirelessDevice', updateWirelessDevice_loRaWAN - The updated wireless device's configuration.

$sel:name:UpdateWirelessDevice', updateWirelessDevice_name - The new name of the resource.

$sel:destinationName:UpdateWirelessDevice', updateWirelessDevice_destinationName - The name of the new destination for the device.

$sel:description:UpdateWirelessDevice', updateWirelessDevice_description - A new description of the resource.

$sel:id:UpdateWirelessDevice', updateWirelessDevice_id - The ID of the resource to update.

Request Lenses

updateWirelessDevice_loRaWAN :: Lens' UpdateWirelessDevice (Maybe LoRaWANUpdateDevice)

The updated wireless device's configuration.

updateWirelessDevice_name :: Lens' UpdateWirelessDevice (Maybe Text)

The new name of the resource.

updateWirelessDevice_destinationName :: Lens' UpdateWirelessDevice (Maybe Text)

The name of the new destination for the device.

updateWirelessDevice_description :: Lens' UpdateWirelessDevice (Maybe Text)

A new description of the resource.

updateWirelessDevice_id :: Lens' UpdateWirelessDevice Text

The ID of the resource to update.

Destructuring the Response

data UpdateWirelessDeviceResponse

See: newUpdateWirelessDeviceResponse smart constructor.

Constructors

UpdateWirelessDeviceResponse'

Fields

httpStatus :: Int

The response's http status code.

newUpdateWirelessDeviceResponse

Arguments

:: Int
$sel:httpStatus:UpdateWirelessDeviceResponse'

-> UpdateWirelessDeviceResponse

Create a value of UpdateWirelessDeviceResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:httpStatus:UpdateWirelessDeviceResponse', updateWirelessDeviceResponse_httpStatus - The response's http status code.

Response Lenses

updateWirelessDeviceResponse_httpStatus :: Lens' UpdateWirelessDeviceResponse Int

The response's http status code.