Amazonka.KMS.ListKeys

Description

Gets a list of all KMS keys in the caller's Amazon Web Services account and Region.

Cross-account use: No. You cannot perform this operation on a KMS key in a different Amazon Web Services account.

Required permissions: kms:ListKeys (IAM policy)

Related operations:

CreateKey

DescribeKey

ListAliases

ListResourceTags

This operation returns paginated results.

Creating a Request

data ListKeys

See: newListKeys smart constructor.

Constructors

ListKeys'

Fields

marker :: Maybe Text

Use this parameter in a subsequent request after you receive a response with truncated results. Set it to the value of NextMarker from the truncated response you just received.

limit :: Maybe Natural

Use this parameter to specify the maximum number of items to return. When this value is present, KMS does not return more than the specified number of items, but it might return fewer.

This value is optional. If you include a value, it must be between 1 and 1000, inclusive. If you do not include a value, it defaults to 100.

newListKeys :: ListKeys

Create a value of ListKeys with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:marker:ListKeys', listKeys_marker - Use this parameter in a subsequent request after you receive a response with truncated results. Set it to the value of NextMarker from the truncated response you just received.

$sel:limit:ListKeys', listKeys_limit - Use this parameter to specify the maximum number of items to return. When this value is present, KMS does not return more than the specified number of items, but it might return fewer.

This value is optional. If you include a value, it must be between 1 and 1000, inclusive. If you do not include a value, it defaults to 100.

Request Lenses

listKeys_marker :: Lens' ListKeys (Maybe Text)

Use this parameter in a subsequent request after you receive a response with truncated results. Set it to the value of NextMarker from the truncated response you just received.

listKeys_limit :: Lens' ListKeys (Maybe Natural)

Use this parameter to specify the maximum number of items to return. When this value is present, KMS does not return more than the specified number of items, but it might return fewer.

This value is optional. If you include a value, it must be between 1 and 1000, inclusive. If you do not include a value, it defaults to 100.

Destructuring the Response

data ListKeysResponse

See: newListKeysResponse smart constructor.

Constructors

ListKeysResponse'

Fields

truncated :: Maybe Bool

A flag that indicates whether there are more items in the list. When this value is true, the list in this response is truncated. To get more items, pass the value of the NextMarker element in thisresponse to the Marker parameter in a subsequent request.

keys :: Maybe [KeyListEntry]

A list of KMS keys.

nextMarker :: Maybe Text

When Truncated is true, this element is present and contains the value to use for the Marker parameter in a subsequent request.

httpStatus :: Int

The response's http status code.

newListKeysResponse

Arguments

:: Int
$sel:httpStatus:ListKeysResponse'

-> ListKeysResponse

Create a value of ListKeysResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:truncated:ListKeysResponse', listKeysResponse_truncated - A flag that indicates whether there are more items in the list. When this value is true, the list in this response is truncated. To get more items, pass the value of the NextMarker element in thisresponse to the Marker parameter in a subsequent request.

$sel:keys:ListKeysResponse', listKeysResponse_keys - A list of KMS keys.

$sel:nextMarker:ListKeysResponse', listKeysResponse_nextMarker - When Truncated is true, this element is present and contains the value to use for the Marker parameter in a subsequent request.

$sel:httpStatus:ListKeysResponse', listKeysResponse_httpStatus - The response's http status code.

Response Lenses

listKeysResponse_truncated :: Lens' ListKeysResponse (Maybe Bool)

A flag that indicates whether there are more items in the list. When this value is true, the list in this response is truncated. To get more items, pass the value of the NextMarker element in thisresponse to the Marker parameter in a subsequent request.

listKeysResponse_keys :: Lens' ListKeysResponse (Maybe [KeyListEntry])

A list of KMS keys.

listKeysResponse_nextMarker :: Lens' ListKeysResponse (Maybe Text)

When Truncated is true, this element is present and contains the value to use for the Marker parameter in a subsequent request.

listKeysResponse_httpStatus :: Lens' ListKeysResponse Int

The response's http status code.