Amazonka.Kafka.GetBootstrapBrokers

Description

A list of brokers that a client application can use to bootstrap.

Creating a Request

data GetBootstrapBrokers

See: newGetBootstrapBrokers smart constructor.

Constructors

GetBootstrapBrokers'

Fields

clusterArn :: Text

The Amazon Resource Name (ARN) that uniquely identifies the cluster.

newGetBootstrapBrokers

Arguments

:: Text
$sel:clusterArn:GetBootstrapBrokers'

-> GetBootstrapBrokers

Create a value of GetBootstrapBrokers with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:clusterArn:GetBootstrapBrokers', getBootstrapBrokers_clusterArn - The Amazon Resource Name (ARN) that uniquely identifies the cluster.

Request Lenses

getBootstrapBrokers_clusterArn :: Lens' GetBootstrapBrokers Text

The Amazon Resource Name (ARN) that uniquely identifies the cluster.

Destructuring the Response

data GetBootstrapBrokersResponse

See: newGetBootstrapBrokersResponse smart constructor.

Constructors

GetBootstrapBrokersResponse'

Fields

bootstrapBrokerString :: Maybe Text

A string containing one or more hostname:port pairs.

bootstrapBrokerStringSaslScram :: Maybe Text

A string containing one or more DNS names (or IP) and Sasl Scram port pairs.

bootstrapBrokerStringTls :: Maybe Text

A string containing one or more DNS names (or IP) and TLS port pairs.

bootstrapBrokerStringSaslIam :: Maybe Text

A string that contains one or more DNS names (or IP addresses) and SASL IAM port pairs.

httpStatus :: Int

The response's http status code.

newGetBootstrapBrokersResponse

Arguments

:: Int
$sel:httpStatus:GetBootstrapBrokersResponse'

-> GetBootstrapBrokersResponse

Create a value of GetBootstrapBrokersResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:bootstrapBrokerString:GetBootstrapBrokersResponse', getBootstrapBrokersResponse_bootstrapBrokerString - A string containing one or more hostname:port pairs.

$sel:bootstrapBrokerStringSaslScram:GetBootstrapBrokersResponse', getBootstrapBrokersResponse_bootstrapBrokerStringSaslScram - A string containing one or more DNS names (or IP) and Sasl Scram port pairs.

$sel:bootstrapBrokerStringTls:GetBootstrapBrokersResponse', getBootstrapBrokersResponse_bootstrapBrokerStringTls - A string containing one or more DNS names (or IP) and TLS port pairs.

$sel:bootstrapBrokerStringSaslIam:GetBootstrapBrokersResponse', getBootstrapBrokersResponse_bootstrapBrokerStringSaslIam - A string that contains one or more DNS names (or IP addresses) and SASL IAM port pairs.

$sel:httpStatus:GetBootstrapBrokersResponse', getBootstrapBrokersResponse_httpStatus - The response's http status code.

Response Lenses

getBootstrapBrokersResponse_bootstrapBrokerString :: Lens' GetBootstrapBrokersResponse (Maybe Text)

A string containing one or more hostname:port pairs.

getBootstrapBrokersResponse_bootstrapBrokerStringSaslScram :: Lens' GetBootstrapBrokersResponse (Maybe Text)

A string containing one or more DNS names (or IP) and Sasl Scram port pairs.

getBootstrapBrokersResponse_bootstrapBrokerStringTls :: Lens' GetBootstrapBrokersResponse (Maybe Text)

A string containing one or more DNS names (or IP) and TLS port pairs.

getBootstrapBrokersResponse_bootstrapBrokerStringSaslIam :: Lens' GetBootstrapBrokersResponse (Maybe Text)

A string that contains one or more DNS names (or IP addresses) and SASL IAM port pairs.

getBootstrapBrokersResponse_httpStatus :: Lens' GetBootstrapBrokersResponse Int

The response's http status code.