Amazonka.Kafka.RebootBroker

Description

Reboots brokers.

Creating a Request

data RebootBroker

Reboots a node.

See: newRebootBroker smart constructor.

Constructors

RebootBroker'

Fields

clusterArn :: Text

The Amazon Resource Name (ARN) of the cluster to be updated.

brokerIds :: [Text]

The list of broker IDs to be rebooted. The reboot-broker operation supports rebooting one broker at a time.

newRebootBroker

Arguments

:: Text
$sel:clusterArn:RebootBroker'

-> RebootBroker

Create a value of RebootBroker with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:clusterArn:RebootBroker', rebootBroker_clusterArn - The Amazon Resource Name (ARN) of the cluster to be updated.

$sel:brokerIds:RebootBroker', rebootBroker_brokerIds - The list of broker IDs to be rebooted. The reboot-broker operation supports rebooting one broker at a time.

Request Lenses

rebootBroker_clusterArn :: Lens' RebootBroker Text

The Amazon Resource Name (ARN) of the cluster to be updated.

rebootBroker_brokerIds :: Lens' RebootBroker [Text]

The list of broker IDs to be rebooted. The reboot-broker operation supports rebooting one broker at a time.

Destructuring the Response

data RebootBrokerResponse

See: newRebootBrokerResponse smart constructor.

Constructors

RebootBrokerResponse'

Fields

clusterArn :: Maybe Text

The Amazon Resource Name (ARN) of the cluster.

clusterOperationArn :: Maybe Text

The Amazon Resource Name (ARN) of the cluster operation.

httpStatus :: Int

The response's http status code.

newRebootBrokerResponse

Arguments

:: Int
$sel:httpStatus:RebootBrokerResponse'

-> RebootBrokerResponse

Create a value of RebootBrokerResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:clusterArn:RebootBroker', rebootBrokerResponse_clusterArn - The Amazon Resource Name (ARN) of the cluster.

$sel:clusterOperationArn:RebootBrokerResponse', rebootBrokerResponse_clusterOperationArn - The Amazon Resource Name (ARN) of the cluster operation.

$sel:httpStatus:RebootBrokerResponse', rebootBrokerResponse_httpStatus - The response's http status code.

Response Lenses

rebootBrokerResponse_clusterArn :: Lens' RebootBrokerResponse (Maybe Text)

The Amazon Resource Name (ARN) of the cluster.

rebootBrokerResponse_clusterOperationArn :: Lens' RebootBrokerResponse (Maybe Text)

The Amazon Resource Name (ARN) of the cluster operation.

rebootBrokerResponse_httpStatus :: Lens' RebootBrokerResponse Int

The response's http status code.