Amazonka.Kafka.Types.ClusterOperationInfo

Description

Documentation

data ClusterOperationInfo

Returns information about a cluster operation.

See: newClusterOperationInfo smart constructor.

Constructors

ClusterOperationInfo'

Fields

creationTime :: Maybe POSIX

The time that the operation was created.

clusterArn :: Maybe Text

ARN of the cluster.

clientRequestId :: Maybe Text

The ID of the API request that triggered this operation.

targetClusterInfo :: Maybe MutableClusterInfo

Information about cluster attributes after a cluster is updated.

sourceClusterInfo :: Maybe MutableClusterInfo

Information about cluster attributes before a cluster is updated.

operationSteps :: Maybe [ClusterOperationStep]

Steps completed during the operation.

operationState :: Maybe Text

State of the cluster operation.

endTime :: Maybe POSIX

The time at which the operation finished.

operationType :: Maybe Text

Type of the cluster operation.

operationArn :: Maybe Text

ARN of the cluster operation.

errorInfo :: Maybe ErrorInfo

Describes the error if the operation fails.

newClusterOperationInfo :: ClusterOperationInfo

Create a value of ClusterOperationInfo with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:creationTime:ClusterOperationInfo', clusterOperationInfo_creationTime - The time that the operation was created.

$sel:clusterArn:ClusterOperationInfo', clusterOperationInfo_clusterArn - ARN of the cluster.

$sel:clientRequestId:ClusterOperationInfo', clusterOperationInfo_clientRequestId - The ID of the API request that triggered this operation.

$sel:targetClusterInfo:ClusterOperationInfo', clusterOperationInfo_targetClusterInfo - Information about cluster attributes after a cluster is updated.

$sel:sourceClusterInfo:ClusterOperationInfo', clusterOperationInfo_sourceClusterInfo - Information about cluster attributes before a cluster is updated.

$sel:operationSteps:ClusterOperationInfo', clusterOperationInfo_operationSteps - Steps completed during the operation.

$sel:operationState:ClusterOperationInfo', clusterOperationInfo_operationState - State of the cluster operation.

$sel:endTime:ClusterOperationInfo', clusterOperationInfo_endTime - The time at which the operation finished.

$sel:operationType:ClusterOperationInfo', clusterOperationInfo_operationType - Type of the cluster operation.

$sel:operationArn:ClusterOperationInfo', clusterOperationInfo_operationArn - ARN of the cluster operation.

$sel:errorInfo:ClusterOperationInfo', clusterOperationInfo_errorInfo - Describes the error if the operation fails.

clusterOperationInfo_creationTime :: Lens' ClusterOperationInfo (Maybe UTCTime)

The time that the operation was created.

clusterOperationInfo_clusterArn :: Lens' ClusterOperationInfo (Maybe Text)

ARN of the cluster.

clusterOperationInfo_clientRequestId :: Lens' ClusterOperationInfo (Maybe Text)

The ID of the API request that triggered this operation.

clusterOperationInfo_targetClusterInfo :: Lens' ClusterOperationInfo (Maybe MutableClusterInfo)

Information about cluster attributes after a cluster is updated.

clusterOperationInfo_sourceClusterInfo :: Lens' ClusterOperationInfo (Maybe MutableClusterInfo)

Information about cluster attributes before a cluster is updated.

clusterOperationInfo_operationSteps :: Lens' ClusterOperationInfo (Maybe [ClusterOperationStep])

Steps completed during the operation.

clusterOperationInfo_operationState :: Lens' ClusterOperationInfo (Maybe Text)

State of the cluster operation.

clusterOperationInfo_endTime :: Lens' ClusterOperationInfo (Maybe UTCTime)

The time at which the operation finished.

clusterOperationInfo_operationType :: Lens' ClusterOperationInfo (Maybe Text)

Type of the cluster operation.

clusterOperationInfo_operationArn :: Lens' ClusterOperationInfo (Maybe Text)

ARN of the cluster operation.

clusterOperationInfo_errorInfo :: Lens' ClusterOperationInfo (Maybe ErrorInfo)

Describes the error if the operation fails.