Amazonka.Kafka.Types.Sasl

Description

Documentation

data Sasl

Details for client authentication using SASL.

See: newSasl smart constructor.

Constructors

Sasl'

Fields

iam :: Maybe Iam

Indicates whether IAM access control is enabled.

scram :: Maybe Scram

Details for SASL/SCRAM client authentication.

newSasl :: Sasl

Create a value of Sasl with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:iam:Sasl', sasl_iam - Indicates whether IAM access control is enabled.

$sel:scram:Sasl', sasl_scram - Details for SASL/SCRAM client authentication.

sasl_iam :: Lens' Sasl (Maybe Iam)

Indicates whether IAM access control is enabled.

sasl_scram :: Lens' Sasl (Maybe Scram)

Details for SASL/SCRAM client authentication.