Amazonka.Kafka.UpdateBrokerType

Description

Updates EC2 instance type.

Creating a Request

data UpdateBrokerType

See: newUpdateBrokerType smart constructor.

Constructors

UpdateBrokerType'

Fields

clusterArn :: Text

The Amazon Resource Name (ARN) that uniquely identifies the cluster.

currentVersion :: Text

The cluster version that you want to change. After this operation completes successfully, the cluster will have a new version.

targetInstanceType :: Text

The Amazon MSK broker type that you want all of the brokers in this cluster to be.

newUpdateBrokerType

Arguments

:: Text
$sel:clusterArn:UpdateBrokerType'

-> Text
$sel:currentVersion:UpdateBrokerType'

-> Text
$sel:targetInstanceType:UpdateBrokerType'

-> UpdateBrokerType

Create a value of UpdateBrokerType with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:clusterArn:UpdateBrokerType', updateBrokerType_clusterArn - The Amazon Resource Name (ARN) that uniquely identifies the cluster.

$sel:currentVersion:UpdateBrokerType', updateBrokerType_currentVersion - The cluster version that you want to change. After this operation completes successfully, the cluster will have a new version.

$sel:targetInstanceType:UpdateBrokerType', updateBrokerType_targetInstanceType - The Amazon MSK broker type that you want all of the brokers in this cluster to be.

Request Lenses

updateBrokerType_clusterArn :: Lens' UpdateBrokerType Text

The Amazon Resource Name (ARN) that uniquely identifies the cluster.

updateBrokerType_currentVersion :: Lens' UpdateBrokerType Text

The cluster version that you want to change. After this operation completes successfully, the cluster will have a new version.

updateBrokerType_targetInstanceType :: Lens' UpdateBrokerType Text

The Amazon MSK broker type that you want all of the brokers in this cluster to be.

Destructuring the Response

data UpdateBrokerTypeResponse

See: newUpdateBrokerTypeResponse smart constructor.

Constructors

UpdateBrokerTypeResponse'

Fields

clusterArn :: Maybe Text

The Amazon Resource Name (ARN) of the cluster.

clusterOperationArn :: Maybe Text

The Amazon Resource Name (ARN) of the cluster operation.

httpStatus :: Int

The response's http status code.

newUpdateBrokerTypeResponse

Arguments

:: Int
$sel:httpStatus:UpdateBrokerTypeResponse'

-> UpdateBrokerTypeResponse

Create a value of UpdateBrokerTypeResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:clusterArn:UpdateBrokerType', updateBrokerTypeResponse_clusterArn - The Amazon Resource Name (ARN) of the cluster.

$sel:clusterOperationArn:UpdateBrokerTypeResponse', updateBrokerTypeResponse_clusterOperationArn - The Amazon Resource Name (ARN) of the cluster operation.

$sel:httpStatus:UpdateBrokerTypeResponse', updateBrokerTypeResponse_httpStatus - The response's http status code.

Response Lenses

updateBrokerTypeResponse_clusterArn :: Lens' UpdateBrokerTypeResponse (Maybe Text)

The Amazon Resource Name (ARN) of the cluster.

updateBrokerTypeResponse_clusterOperationArn :: Lens' UpdateBrokerTypeResponse (Maybe Text)

The Amazon Resource Name (ARN) of the cluster operation.

updateBrokerTypeResponse_httpStatus :: Lens' UpdateBrokerTypeResponse Int

The response's http status code.