Amazonka.KafkaConnect.DescribeConnector

Description

Returns summary information about the connector.

Creating a Request

data DescribeConnector

See: newDescribeConnector smart constructor.

Constructors

DescribeConnector'

Fields

connectorArn :: Text

The Amazon Resource Name (ARN) of the connector that you want to describe.

newDescribeConnector

Arguments

:: Text
$sel:connectorArn:DescribeConnector'

-> DescribeConnector

Create a value of DescribeConnector with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:connectorArn:DescribeConnector', describeConnector_connectorArn - The Amazon Resource Name (ARN) of the connector that you want to describe.

Request Lenses

describeConnector_connectorArn :: Lens' DescribeConnector Text

The Amazon Resource Name (ARN) of the connector that you want to describe.

Destructuring the Response

data DescribeConnectorResponse

See: newDescribeConnectorResponse smart constructor.

Constructors

DescribeConnectorResponse'

Fields

creationTime :: Maybe POSIX

The time the connector was created.

kafkaCluster :: Maybe KafkaClusterDescription

The Apache Kafka cluster that the connector is connected to.

kafkaConnectVersion :: Maybe Text

The version of Kafka Connect. It has to be compatible with both the Apache Kafka cluster's version and the plugins.

logDelivery :: Maybe LogDeliveryDescription

Details about delivering logs to Amazon CloudWatch Logs.

currentVersion :: Maybe Text

The current version of the connector.

connectorConfiguration :: Maybe (HashMap Text Text)

A map of keys to values that represent the configuration for the connector.

workerConfiguration :: Maybe WorkerConfigurationDescription

Specifies which worker configuration was used for the connector.

connectorArn :: Maybe Text

The Amazon Resource Name (ARN) of the connector.

connectorName :: Maybe Text

The name of the connector.

connectorState :: Maybe ConnectorState

The state of the connector.

capacity :: Maybe CapacityDescription

Information about the capacity of the connector, whether it is auto scaled or provisioned.

plugins :: Maybe [PluginDescription]

Specifies which plugins were used for this connector.

connectorDescription :: Maybe Text

A summary description of the connector.

kafkaClusterClientAuthentication :: Maybe KafkaClusterClientAuthenticationDescription

The type of client authentication used to connect to the Apache Kafka cluster. The value is NONE when no client authentication is used.

kafkaClusterEncryptionInTransit :: Maybe KafkaClusterEncryptionInTransitDescription

Details of encryption in transit to the Apache Kafka cluster.

serviceExecutionRoleArn :: Maybe Text

The Amazon Resource Name (ARN) of the IAM role used by the connector to access Amazon Web Services resources.

httpStatus :: Int

The response's http status code.

newDescribeConnectorResponse

Arguments

:: Int
$sel:httpStatus:DescribeConnectorResponse'

-> DescribeConnectorResponse

Create a value of DescribeConnectorResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:creationTime:DescribeConnectorResponse', describeConnectorResponse_creationTime - The time the connector was created.

$sel:kafkaCluster:DescribeConnectorResponse', describeConnectorResponse_kafkaCluster - The Apache Kafka cluster that the connector is connected to.

$sel:kafkaConnectVersion:DescribeConnectorResponse', describeConnectorResponse_kafkaConnectVersion - The version of Kafka Connect. It has to be compatible with both the Apache Kafka cluster's version and the plugins.

$sel:logDelivery:DescribeConnectorResponse', describeConnectorResponse_logDelivery - Details about delivering logs to Amazon CloudWatch Logs.

$sel:currentVersion:DescribeConnectorResponse', describeConnectorResponse_currentVersion - The current version of the connector.

$sel:connectorConfiguration:DescribeConnectorResponse', describeConnectorResponse_connectorConfiguration - A map of keys to values that represent the configuration for the connector.

$sel:workerConfiguration:DescribeConnectorResponse', describeConnectorResponse_workerConfiguration - Specifies which worker configuration was used for the connector.

$sel:connectorArn:DescribeConnector', describeConnectorResponse_connectorArn - The Amazon Resource Name (ARN) of the connector.

$sel:connectorName:DescribeConnectorResponse', describeConnectorResponse_connectorName - The name of the connector.

$sel:connectorState:DescribeConnectorResponse', describeConnectorResponse_connectorState - The state of the connector.

$sel:capacity:DescribeConnectorResponse', describeConnectorResponse_capacity - Information about the capacity of the connector, whether it is auto scaled or provisioned.

$sel:plugins:DescribeConnectorResponse', describeConnectorResponse_plugins - Specifies which plugins were used for this connector.

$sel:connectorDescription:DescribeConnectorResponse', describeConnectorResponse_connectorDescription - A summary description of the connector.

$sel:kafkaClusterClientAuthentication:DescribeConnectorResponse', describeConnectorResponse_kafkaClusterClientAuthentication - The type of client authentication used to connect to the Apache Kafka cluster. The value is NONE when no client authentication is used.

$sel:kafkaClusterEncryptionInTransit:DescribeConnectorResponse', describeConnectorResponse_kafkaClusterEncryptionInTransit - Details of encryption in transit to the Apache Kafka cluster.

$sel:serviceExecutionRoleArn:DescribeConnectorResponse', describeConnectorResponse_serviceExecutionRoleArn - The Amazon Resource Name (ARN) of the IAM role used by the connector to access Amazon Web Services resources.

$sel:httpStatus:DescribeConnectorResponse', describeConnectorResponse_httpStatus - The response's http status code.

Response Lenses

describeConnectorResponse_creationTime :: Lens' DescribeConnectorResponse (Maybe UTCTime)

The time the connector was created.

describeConnectorResponse_kafkaCluster :: Lens' DescribeConnectorResponse (Maybe KafkaClusterDescription)

The Apache Kafka cluster that the connector is connected to.

describeConnectorResponse_kafkaConnectVersion :: Lens' DescribeConnectorResponse (Maybe Text)

The version of Kafka Connect. It has to be compatible with both the Apache Kafka cluster's version and the plugins.

describeConnectorResponse_logDelivery :: Lens' DescribeConnectorResponse (Maybe LogDeliveryDescription)

Details about delivering logs to Amazon CloudWatch Logs.

describeConnectorResponse_currentVersion :: Lens' DescribeConnectorResponse (Maybe Text)

The current version of the connector.

describeConnectorResponse_connectorConfiguration :: Lens' DescribeConnectorResponse (Maybe (HashMap Text Text))

A map of keys to values that represent the configuration for the connector.

describeConnectorResponse_workerConfiguration :: Lens' DescribeConnectorResponse (Maybe WorkerConfigurationDescription)

Specifies which worker configuration was used for the connector.

describeConnectorResponse_connectorArn :: Lens' DescribeConnectorResponse (Maybe Text)

The Amazon Resource Name (ARN) of the connector.

describeConnectorResponse_connectorName :: Lens' DescribeConnectorResponse (Maybe Text)

The name of the connector.

describeConnectorResponse_connectorState :: Lens' DescribeConnectorResponse (Maybe ConnectorState)

The state of the connector.

describeConnectorResponse_capacity :: Lens' DescribeConnectorResponse (Maybe CapacityDescription)

Information about the capacity of the connector, whether it is auto scaled or provisioned.

describeConnectorResponse_plugins :: Lens' DescribeConnectorResponse (Maybe [PluginDescription])

Specifies which plugins were used for this connector.

describeConnectorResponse_connectorDescription :: Lens' DescribeConnectorResponse (Maybe Text)

A summary description of the connector.

describeConnectorResponse_kafkaClusterClientAuthentication :: Lens' DescribeConnectorResponse (Maybe KafkaClusterClientAuthenticationDescription)

The type of client authentication used to connect to the Apache Kafka cluster. The value is NONE when no client authentication is used.

describeConnectorResponse_kafkaClusterEncryptionInTransit :: Lens' DescribeConnectorResponse (Maybe KafkaClusterEncryptionInTransitDescription)

Details of encryption in transit to the Apache Kafka cluster.

describeConnectorResponse_serviceExecutionRoleArn :: Lens' DescribeConnectorResponse (Maybe Text)

The Amazon Resource Name (ARN) of the IAM role used by the connector to access Amazon Web Services resources.

describeConnectorResponse_httpStatus :: Lens' DescribeConnectorResponse Int

The response's http status code.