Amazonka.KafkaConnect.ListCustomPlugins

Description

Returns a list of all of the custom plugins in this account and Region.

This operation returns paginated results.

Creating a Request

data ListCustomPlugins

See: newListCustomPlugins smart constructor.

Constructors

ListCustomPlugins'

Fields

nextToken :: Maybe Text

If the response of a ListCustomPlugins operation is truncated, it will include a NextToken. Send this NextToken in a subsequent request to continue listing from where the previous operation left off.

maxResults :: Maybe Natural

The maximum number of custom plugins to list in one response.

newListCustomPlugins :: ListCustomPlugins

Create a value of ListCustomPlugins with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:nextToken:ListCustomPlugins', listCustomPlugins_nextToken - If the response of a ListCustomPlugins operation is truncated, it will include a NextToken. Send this NextToken in a subsequent request to continue listing from where the previous operation left off.

$sel:maxResults:ListCustomPlugins', listCustomPlugins_maxResults - The maximum number of custom plugins to list in one response.

Request Lenses

listCustomPlugins_nextToken :: Lens' ListCustomPlugins (Maybe Text)

If the response of a ListCustomPlugins operation is truncated, it will include a NextToken. Send this NextToken in a subsequent request to continue listing from where the previous operation left off.

listCustomPlugins_maxResults :: Lens' ListCustomPlugins (Maybe Natural)

The maximum number of custom plugins to list in one response.

Destructuring the Response

data ListCustomPluginsResponse

See: newListCustomPluginsResponse smart constructor.

Constructors

ListCustomPluginsResponse'

Fields

customPlugins :: Maybe [CustomPluginSummary]

An array of custom plugin descriptions.

nextToken :: Maybe Text

If the response of a ListCustomPlugins operation is truncated, it will include a NextToken. Send this NextToken in a subsequent request to continue listing from where the previous operation left off.

httpStatus :: Int

The response's http status code.

newListCustomPluginsResponse

Arguments

:: Int
$sel:httpStatus:ListCustomPluginsResponse'

-> ListCustomPluginsResponse

Create a value of ListCustomPluginsResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:customPlugins:ListCustomPluginsResponse', listCustomPluginsResponse_customPlugins - An array of custom plugin descriptions.

$sel:nextToken:ListCustomPlugins', listCustomPluginsResponse_nextToken - If the response of a ListCustomPlugins operation is truncated, it will include a NextToken. Send this NextToken in a subsequent request to continue listing from where the previous operation left off.

$sel:httpStatus:ListCustomPluginsResponse', listCustomPluginsResponse_httpStatus - The response's http status code.

Response Lenses

listCustomPluginsResponse_customPlugins :: Lens' ListCustomPluginsResponse (Maybe [CustomPluginSummary])

An array of custom plugin descriptions.

listCustomPluginsResponse_nextToken :: Lens' ListCustomPluginsResponse (Maybe Text)

If the response of a ListCustomPlugins operation is truncated, it will include a NextToken. Send this NextToken in a subsequent request to continue listing from where the previous operation left off.

listCustomPluginsResponse_httpStatus :: Lens' ListCustomPluginsResponse Int

The response's http status code.