Amazonka.KafkaConnect.Types.AutoScalingDescription

Description

Documentation

data AutoScalingDescription

Information about the auto scaling parameters for the connector.

See: newAutoScalingDescription smart constructor.

Constructors

AutoScalingDescription'

Fields

scaleInPolicy :: Maybe ScaleInPolicyDescription

The sacle-in policy for the connector.

mcuCount :: Maybe Int

The number of microcontroller units (MCUs) allocated to each connector worker. The valid values are 1,2,4,8.

scaleOutPolicy :: Maybe ScaleOutPolicyDescription

The sacle-out policy for the connector.>

maxWorkerCount :: Maybe Int

The maximum number of workers allocated to the connector.

minWorkerCount :: Maybe Int

The minimum number of workers allocated to the connector.

newAutoScalingDescription :: AutoScalingDescription

Create a value of AutoScalingDescription with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:scaleInPolicy:AutoScalingDescription', autoScalingDescription_scaleInPolicy - The sacle-in policy for the connector.

$sel:mcuCount:AutoScalingDescription', autoScalingDescription_mcuCount - The number of microcontroller units (MCUs) allocated to each connector worker. The valid values are 1,2,4,8.

$sel:scaleOutPolicy:AutoScalingDescription', autoScalingDescription_scaleOutPolicy - The sacle-out policy for the connector.>

$sel:maxWorkerCount:AutoScalingDescription', autoScalingDescription_maxWorkerCount - The maximum number of workers allocated to the connector.

$sel:minWorkerCount:AutoScalingDescription', autoScalingDescription_minWorkerCount - The minimum number of workers allocated to the connector.

autoScalingDescription_scaleInPolicy :: Lens' AutoScalingDescription (Maybe ScaleInPolicyDescription)

The sacle-in policy for the connector.

autoScalingDescription_mcuCount :: Lens' AutoScalingDescription (Maybe Int)

The number of microcontroller units (MCUs) allocated to each connector worker. The valid values are 1,2,4,8.

autoScalingDescription_scaleOutPolicy :: Lens' AutoScalingDescription (Maybe ScaleOutPolicyDescription)

The sacle-out policy for the connector.>

autoScalingDescription_maxWorkerCount :: Lens' AutoScalingDescription (Maybe Int)

The maximum number of workers allocated to the connector.

autoScalingDescription_minWorkerCount :: Lens' AutoScalingDescription (Maybe Int)

The minimum number of workers allocated to the connector.