Amazonka.KafkaConnect.Types.AutoScalingUpdate

Description

Documentation

data AutoScalingUpdate

The updates to the auto scaling parameters for the connector.

See: newAutoScalingUpdate smart constructor.

Constructors

AutoScalingUpdate'

Fields

maxWorkerCount :: Natural

The target maximum number of workers allocated to the connector.

mcuCount :: Natural

The target number of microcontroller units (MCUs) allocated to each connector worker. The valid values are 1,2,4,8.

minWorkerCount :: Natural

The target minimum number of workers allocated to the connector.

scaleInPolicy :: ScaleInPolicyUpdate

The target sacle-in policy for the connector.

scaleOutPolicy :: ScaleOutPolicyUpdate

The target sacle-out policy for the connector.

newAutoScalingUpdate

Arguments

:: Natural
$sel:maxWorkerCount:AutoScalingUpdate'

-> Natural
$sel:mcuCount:AutoScalingUpdate'

-> Natural
$sel:minWorkerCount:AutoScalingUpdate'

-> ScaleInPolicyUpdate
$sel:scaleInPolicy:AutoScalingUpdate'

-> ScaleOutPolicyUpdate
$sel:scaleOutPolicy:AutoScalingUpdate'

-> AutoScalingUpdate

Create a value of AutoScalingUpdate with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:maxWorkerCount:AutoScalingUpdate', autoScalingUpdate_maxWorkerCount - The target maximum number of workers allocated to the connector.

$sel:mcuCount:AutoScalingUpdate', autoScalingUpdate_mcuCount - The target number of microcontroller units (MCUs) allocated to each connector worker. The valid values are 1,2,4,8.

$sel:minWorkerCount:AutoScalingUpdate', autoScalingUpdate_minWorkerCount - The target minimum number of workers allocated to the connector.

$sel:scaleInPolicy:AutoScalingUpdate', autoScalingUpdate_scaleInPolicy - The target sacle-in policy for the connector.

$sel:scaleOutPolicy:AutoScalingUpdate', autoScalingUpdate_scaleOutPolicy - The target sacle-out policy for the connector.

autoScalingUpdate_maxWorkerCount :: Lens' AutoScalingUpdate Natural

The target maximum number of workers allocated to the connector.

autoScalingUpdate_mcuCount :: Lens' AutoScalingUpdate Natural

The target number of microcontroller units (MCUs) allocated to each connector worker. The valid values are 1,2,4,8.

autoScalingUpdate_minWorkerCount :: Lens' AutoScalingUpdate Natural

The target minimum number of workers allocated to the connector.

autoScalingUpdate_scaleInPolicy :: Lens' AutoScalingUpdate ScaleInPolicyUpdate

The target sacle-in policy for the connector.

autoScalingUpdate_scaleOutPolicy :: Lens' AutoScalingUpdate ScaleOutPolicyUpdate

The target sacle-out policy for the connector.