Amazonka.KafkaConnect.Types.CustomPluginLocationDescription

Description

Documentation

data CustomPluginLocationDescription

Information about the location of a custom plugin.

See: newCustomPluginLocationDescription smart constructor.

Constructors

CustomPluginLocationDescription'

Fields

s3Location :: Maybe S3LocationDescription

The S3 bucket Amazon Resource Name (ARN), file key, and object version of the plugin file stored in Amazon S3.

newCustomPluginLocationDescription :: CustomPluginLocationDescription

Create a value of CustomPluginLocationDescription with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:s3Location:CustomPluginLocationDescription', customPluginLocationDescription_s3Location - The S3 bucket Amazon Resource Name (ARN), file key, and object version of the plugin file stored in Amazon S3.

customPluginLocationDescription_s3Location :: Lens' CustomPluginLocationDescription (Maybe S3LocationDescription)

The S3 bucket Amazon Resource Name (ARN), file key, and object version of the plugin file stored in Amazon S3.