Amazonka.Kinesis.EnableEnhancedMonitoring

Description

Enables enhanced Kinesis data stream monitoring for shard-level metrics.

Creating a Request

data EnableEnhancedMonitoring

Represents the input for EnableEnhancedMonitoring.

See: newEnableEnhancedMonitoring smart constructor.

Constructors

EnableEnhancedMonitoring'

Fields

streamName :: Text

The name of the stream for which to enable enhanced monitoring.

shardLevelMetrics :: [MetricsName]

List of shard-level metrics to enable.

The following are the valid shard-level metrics. The value "ALL" enables every metric.

IncomingBytes

IncomingRecords

OutgoingBytes

OutgoingRecords

WriteProvisionedThroughputExceeded

ReadProvisionedThroughputExceeded

IteratorAgeMilliseconds

ALL

For more information, see Monitoring the Amazon Kinesis Data Streams Service with Amazon CloudWatch in the Amazon Kinesis Data Streams Developer Guide.

newEnableEnhancedMonitoring

Arguments

:: Text
$sel:streamName:EnableEnhancedMonitoring'

-> EnableEnhancedMonitoring

Create a value of EnableEnhancedMonitoring with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:streamName:EnableEnhancedMonitoring', enableEnhancedMonitoring_streamName - The name of the stream for which to enable enhanced monitoring.

$sel:shardLevelMetrics:EnableEnhancedMonitoring', enableEnhancedMonitoring_shardLevelMetrics - List of shard-level metrics to enable.

The following are the valid shard-level metrics. The value "ALL" enables every metric.

IncomingBytes

IncomingRecords

OutgoingBytes

OutgoingRecords

WriteProvisionedThroughputExceeded

ReadProvisionedThroughputExceeded

IteratorAgeMilliseconds

ALL

For more information, see Monitoring the Amazon Kinesis Data Streams Service with Amazon CloudWatch in the Amazon Kinesis Data Streams Developer Guide.

Request Lenses

enableEnhancedMonitoring_streamName :: Lens' EnableEnhancedMonitoring Text

The name of the stream for which to enable enhanced monitoring.

enableEnhancedMonitoring_shardLevelMetrics :: Lens' EnableEnhancedMonitoring [MetricsName]

List of shard-level metrics to enable.

The following are the valid shard-level metrics. The value "ALL" enables every metric.

IncomingBytes

IncomingRecords

OutgoingBytes

OutgoingRecords

WriteProvisionedThroughputExceeded

ReadProvisionedThroughputExceeded

IteratorAgeMilliseconds

ALL

For more information, see Monitoring the Amazon Kinesis Data Streams Service with Amazon CloudWatch in the Amazon Kinesis Data Streams Developer Guide.

Destructuring the Response

data EnhancedMonitoringOutput

Represents the output for EnableEnhancedMonitoring and DisableEnhancedMonitoring.

See: newEnhancedMonitoringOutput smart constructor.

Constructors

EnhancedMonitoringOutput'

Fields

desiredShardLevelMetrics :: Maybe [MetricsName]

Represents the list of all the metrics that would be in the enhanced state after the operation.

currentShardLevelMetrics :: Maybe [MetricsName]

Represents the current state of the metrics that are in the enhanced state before the operation.

streamName :: Maybe Text

The name of the Kinesis data stream.

newEnhancedMonitoringOutput :: EnhancedMonitoringOutput

Create a value of EnhancedMonitoringOutput with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:desiredShardLevelMetrics:EnhancedMonitoringOutput', enhancedMonitoringOutput_desiredShardLevelMetrics - Represents the list of all the metrics that would be in the enhanced state after the operation.

$sel:currentShardLevelMetrics:EnhancedMonitoringOutput', enhancedMonitoringOutput_currentShardLevelMetrics - Represents the current state of the metrics that are in the enhanced state before the operation.

$sel:streamName:EnhancedMonitoringOutput', enhancedMonitoringOutput_streamName - The name of the Kinesis data stream.

Response Lenses

enhancedMonitoringOutput_desiredShardLevelMetrics :: Lens' EnhancedMonitoringOutput (Maybe [MetricsName])

Represents the list of all the metrics that would be in the enhanced state after the operation.

enhancedMonitoringOutput_currentShardLevelMetrics :: Lens' EnhancedMonitoringOutput (Maybe [MetricsName])

Represents the current state of the metrics that are in the enhanced state before the operation.

enhancedMonitoringOutput_streamName :: Lens' EnhancedMonitoringOutput (Maybe Text)

The name of the Kinesis data stream.