Amazonka.Kinesis.ListTagsForStream

Description

Lists the tags for the specified Kinesis data stream. This operation has a limit of five transactions per second per account.

Creating a Request

data ListTagsForStream

Represents the input for ListTagsForStream.

See: newListTagsForStream smart constructor.

Constructors

ListTagsForStream'

Fields

limit :: Maybe Natural

The number of tags to return. If this number is less than the total number of tags associated with the stream, HasMoreTags is set to true. To list additional tags, set ExclusiveStartTagKey to the last key in the response.

exclusiveStartTagKey :: Maybe Text

The key to use as the starting point for the list of tags. If this parameter is set, ListTagsForStream gets all tags that occur after ExclusiveStartTagKey.

streamName :: Text

The name of the stream.

newListTagsForStream

Arguments

:: Text
$sel:streamName:ListTagsForStream'

-> ListTagsForStream

Create a value of ListTagsForStream with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:limit:ListTagsForStream', listTagsForStream_limit - The number of tags to return. If this number is less than the total number of tags associated with the stream, HasMoreTags is set to true. To list additional tags, set ExclusiveStartTagKey to the last key in the response.

$sel:exclusiveStartTagKey:ListTagsForStream', listTagsForStream_exclusiveStartTagKey - The key to use as the starting point for the list of tags. If this parameter is set, ListTagsForStream gets all tags that occur after ExclusiveStartTagKey.

$sel:streamName:ListTagsForStream', listTagsForStream_streamName - The name of the stream.

Request Lenses

listTagsForStream_limit :: Lens' ListTagsForStream (Maybe Natural)

The number of tags to return. If this number is less than the total number of tags associated with the stream, HasMoreTags is set to true. To list additional tags, set ExclusiveStartTagKey to the last key in the response.

listTagsForStream_exclusiveStartTagKey :: Lens' ListTagsForStream (Maybe Text)

The key to use as the starting point for the list of tags. If this parameter is set, ListTagsForStream gets all tags that occur after ExclusiveStartTagKey.

listTagsForStream_streamName :: Lens' ListTagsForStream Text

The name of the stream.

Destructuring the Response

data ListTagsForStreamResponse

Represents the output for ListTagsForStream.

See: newListTagsForStreamResponse smart constructor.

Constructors

ListTagsForStreamResponse'

Fields

httpStatus :: Int

The response's http status code.

tags :: [Tag]

A list of tags associated with StreamName, starting with the first tag after ExclusiveStartTagKey and up to the specified Limit.

hasMoreTags :: Bool

If set to true, more tags are available. To request additional tags, set ExclusiveStartTagKey to the key of the last tag returned.

newListTagsForStreamResponse

Arguments

:: Int
$sel:httpStatus:ListTagsForStreamResponse'

-> Bool
$sel:hasMoreTags:ListTagsForStreamResponse'

-> ListTagsForStreamResponse

Create a value of ListTagsForStreamResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:httpStatus:ListTagsForStreamResponse', listTagsForStreamResponse_httpStatus - The response's http status code.

$sel:tags:ListTagsForStreamResponse', listTagsForStreamResponse_tags - A list of tags associated with StreamName, starting with the first tag after ExclusiveStartTagKey and up to the specified Limit.

$sel:hasMoreTags:ListTagsForStreamResponse', listTagsForStreamResponse_hasMoreTags - If set to true, more tags are available. To request additional tags, set ExclusiveStartTagKey to the key of the last tag returned.

Response Lenses

listTagsForStreamResponse_httpStatus :: Lens' ListTagsForStreamResponse Int

The response's http status code.

listTagsForStreamResponse_tags :: Lens' ListTagsForStreamResponse [Tag]

A list of tags associated with StreamName, starting with the first tag after ExclusiveStartTagKey and up to the specified Limit.

listTagsForStreamResponse_hasMoreTags :: Lens' ListTagsForStreamResponse Bool

If set to true, more tags are available. To request additional tags, set ExclusiveStartTagKey to the key of the last tag returned.