Amazonka.Kinesis.PutRecords

Description

Writes multiple data records into a Kinesis data stream in a single call (also referred to as a PutRecords request). Use this operation to send data into the stream for data ingestion and processing.

Each PutRecords request can support up to 500 records. Each record in the request can be as large as 1 MiB, up to a limit of 5 MiB for the entire request, including partition keys. Each shard can support writes up to 1,000 records per second, up to a maximum data write total of 1 MiB per second.

You must specify the name of the stream that captures, stores, and transports the data; and an array of request Records, with each record in the array requiring a partition key and data blob. The record size limit applies to the total size of the partition key and data blob.

The data blob can be any type of data; for example, a segment from a log file, geographic/location data, website clickstream data, and so on.

The partition key is used by Kinesis Data Streams as input to a hash function that maps the partition key and associated data to a specific shard. An MD5 hash function is used to map partition keys to 128-bit integer values and to map associated data records to shards. As a result of this hashing mechanism, all data records with the same partition key map to the same shard within the stream. For more information, see Adding Data to a Stream in the Amazon Kinesis Data Streams Developer Guide.

Each record in the Records array may include an optional parameter, ExplicitHashKey, which overrides the partition key to shard mapping. This parameter allows a data producer to determine explicitly the shard where the record is stored. For more information, see Adding Multiple Records with PutRecords in the Amazon Kinesis Data Streams Developer Guide.

The PutRecords response includes an array of response Records. Each record in the response array directly correlates with a record in the request array using natural ordering, from the top to the bottom of the request and response. The response Records array always includes the same number of records as the request array.

The response Records array includes both successfully and unsuccessfully processed records. Kinesis Data Streams attempts to process all records in each PutRecords request. A single record failure does not stop the processing of subsequent records. As a result, PutRecords doesn't guarantee the ordering of records. If you need to read records in the same order they are written to the stream, use PutRecord instead of PutRecords, and write to the same shard.

A successfully processed record includes ShardId and SequenceNumber values. The ShardId parameter identifies the shard in the stream where the record is stored. The SequenceNumber parameter is an identifier assigned to the put record, unique to all records in the stream.

An unsuccessfully processed record includes ErrorCode and ErrorMessage values. ErrorCode reflects the type of error and can be one of the following values: ProvisionedThroughputExceededException or InternalFailure. ErrorMessage provides more detailed information about the ProvisionedThroughputExceededException exception including the account ID, stream name, and shard ID of the record that was throttled. For more information about partially successful responses, see Adding Multiple Records with PutRecords in the Amazon Kinesis Data Streams Developer Guide.

After you write a record to a stream, you cannot modify that record or its order within the stream.

By default, data records are accessible for 24 hours from the time that they are added to a stream. You can use IncreaseStreamRetentionPeriod or DecreaseStreamRetentionPeriod to modify this retention period.

Creating a Request

data PutRecords

A PutRecords request.

See: newPutRecords smart constructor.

Constructors

PutRecords'

Fields

records :: NonEmpty PutRecordsRequestEntry

The records associated with the request.

streamName :: Text

The stream name associated with the request.

newPutRecords

Arguments

:: NonEmpty PutRecordsRequestEntry
$sel:records:PutRecords'

-> Text
$sel:streamName:PutRecords'

-> PutRecords

Create a value of PutRecords with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:records:PutRecords', putRecords_records - The records associated with the request.

$sel:streamName:PutRecords', putRecords_streamName - The stream name associated with the request.

Request Lenses

putRecords_records :: Lens' PutRecords (NonEmpty PutRecordsRequestEntry)

The records associated with the request.

putRecords_streamName :: Lens' PutRecords Text

The stream name associated with the request.

Destructuring the Response

data PutRecordsResponse

PutRecords results.

See: newPutRecordsResponse smart constructor.

Constructors

PutRecordsResponse'

Fields

encryptionType :: Maybe EncryptionType

The encryption type used on the records. This parameter can be one of the following values:

NONE: Do not encrypt the records.

KMS: Use server-side encryption on the records using a customer-managed AWS KMS key.

failedRecordCount :: Maybe Natural

The number of unsuccessfully processed records in a PutRecords request.

httpStatus :: Int

The response's http status code.

records :: NonEmpty PutRecordsResultEntry

An array of successfully and unsuccessfully processed record results, correlated with the request by natural ordering. A record that is successfully added to a stream includes SequenceNumber and ShardId in the result. A record that fails to be added to a stream includes ErrorCode and ErrorMessage in the result.

newPutRecordsResponse

Arguments

:: Int
$sel:httpStatus:PutRecordsResponse'

-> NonEmpty PutRecordsResultEntry
$sel:records:PutRecords'

-> PutRecordsResponse

Create a value of PutRecordsResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:encryptionType:PutRecordsResponse', putRecordsResponse_encryptionType - The encryption type used on the records. This parameter can be one of the following values:

NONE: Do not encrypt the records.

KMS: Use server-side encryption on the records using a customer-managed AWS KMS key.

$sel:failedRecordCount:PutRecordsResponse', putRecordsResponse_failedRecordCount - The number of unsuccessfully processed records in a PutRecords request.

$sel:httpStatus:PutRecordsResponse', putRecordsResponse_httpStatus - The response's http status code.

$sel:records:PutRecords', putRecordsResponse_records - An array of successfully and unsuccessfully processed record results, correlated with the request by natural ordering. A record that is successfully added to a stream includes SequenceNumber and ShardId in the result. A record that fails to be added to a stream includes ErrorCode and ErrorMessage in the result.

Response Lenses

putRecordsResponse_encryptionType :: Lens' PutRecordsResponse (Maybe EncryptionType)

The encryption type used on the records. This parameter can be one of the following values:

NONE: Do not encrypt the records.

KMS: Use server-side encryption on the records using a customer-managed AWS KMS key.

putRecordsResponse_failedRecordCount :: Lens' PutRecordsResponse (Maybe Natural)

The number of unsuccessfully processed records in a PutRecords request.

putRecordsResponse_httpStatus :: Lens' PutRecordsResponse Int

The response's http status code.

putRecordsResponse_records :: Lens' PutRecordsResponse (NonEmpty PutRecordsResultEntry)

An array of successfully and unsuccessfully processed record results, correlated with the request by natural ordering. A record that is successfully added to a stream includes SequenceNumber and ShardId in the result. A record that fails to be added to a stream includes ErrorCode and ErrorMessage in the result.