Amazonka.Kinesis.Types.Consumer

Description

Documentation

data Consumer

An object that represents the details of the consumer you registered. This type of object is returned by RegisterStreamConsumer.

See: newConsumer smart constructor.

Constructors

Consumer'

Fields

consumerName :: Text

The name of the consumer is something you choose when you register the consumer.

consumerARN :: Text

When you register a consumer, Kinesis Data Streams generates an ARN for it. You need this ARN to be able to call SubscribeToShard.

If you delete a consumer and then create a new one with the same name, it won't have the same ARN. That's because consumer ARNs contain the creation timestamp. This is important to keep in mind if you have IAM policies that reference consumer ARNs.

consumerStatus :: ConsumerStatus

A consumer can't read data while in the CREATING or DELETING states.

consumerCreationTimestamp :: POSIX

newConsumer

Arguments

:: Text
$sel:consumerName:Consumer'

-> Text
$sel:consumerARN:Consumer'

-> ConsumerStatus
$sel:consumerStatus:Consumer'

-> UTCTime
$sel:consumerCreationTimestamp:Consumer'

-> Consumer

Create a value of Consumer with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:consumerName:Consumer', consumer_consumerName - The name of the consumer is something you choose when you register the consumer.

$sel:consumerARN:Consumer', consumer_consumerARN - When you register a consumer, Kinesis Data Streams generates an ARN for it. You need this ARN to be able to call SubscribeToShard.

If you delete a consumer and then create a new one with the same name, it won't have the same ARN. That's because consumer ARNs contain the creation timestamp. This is important to keep in mind if you have IAM policies that reference consumer ARNs.

$sel:consumerStatus:Consumer', consumer_consumerStatus - A consumer can't read data while in the CREATING or DELETING states.

$sel:consumerCreationTimestamp:Consumer', consumer_consumerCreationTimestamp -

consumer_consumerName :: Lens' Consumer Text

The name of the consumer is something you choose when you register the consumer.

consumer_consumerARN :: Lens' Consumer Text

When you register a consumer, Kinesis Data Streams generates an ARN for it. You need this ARN to be able to call SubscribeToShard.

If you delete a consumer and then create a new one with the same name, it won't have the same ARN. That's because consumer ARNs contain the creation timestamp. This is important to keep in mind if you have IAM policies that reference consumer ARNs.

consumer_consumerStatus :: Lens' Consumer ConsumerStatus

A consumer can't read data while in the CREATING or DELETING states.

consumer_consumerCreationTimestamp :: Lens' Consumer UTCTime