Amazonka.Kinesis.Types.EnhancedMetrics

Description

Documentation

data EnhancedMetrics

Represents enhanced metrics types.

See: newEnhancedMetrics smart constructor.

Constructors

EnhancedMetrics'

Fields

shardLevelMetrics :: Maybe [MetricsName]

List of shard-level metrics.

The following are the valid shard-level metrics. The value "ALL" enhances every metric.

IncomingBytes

IncomingRecords

OutgoingBytes

OutgoingRecords

WriteProvisionedThroughputExceeded

ReadProvisionedThroughputExceeded

IteratorAgeMilliseconds

ALL

For more information, see Monitoring the Amazon Kinesis Data Streams Service with Amazon CloudWatch in the Amazon Kinesis Data Streams Developer Guide.

newEnhancedMetrics :: EnhancedMetrics

Create a value of EnhancedMetrics with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:shardLevelMetrics:EnhancedMetrics', enhancedMetrics_shardLevelMetrics - List of shard-level metrics.

The following are the valid shard-level metrics. The value "ALL" enhances every metric.

IncomingBytes

IncomingRecords

OutgoingBytes

OutgoingRecords

WriteProvisionedThroughputExceeded

ReadProvisionedThroughputExceeded

IteratorAgeMilliseconds

ALL

For more information, see Monitoring the Amazon Kinesis Data Streams Service with Amazon CloudWatch in the Amazon Kinesis Data Streams Developer Guide.

enhancedMetrics_shardLevelMetrics :: Lens' EnhancedMetrics (Maybe [MetricsName])

List of shard-level metrics.

The following are the valid shard-level metrics. The value "ALL" enhances every metric.

IncomingBytes

IncomingRecords

OutgoingBytes

OutgoingRecords

WriteProvisionedThroughputExceeded

ReadProvisionedThroughputExceeded

IteratorAgeMilliseconds

ALL

For more information, see Monitoring the Amazon Kinesis Data Streams Service with Amazon CloudWatch in the Amazon Kinesis Data Streams Developer Guide.