Amazonka.Kinesis.Types.StreamDescriptionSummary

Description

Documentation

data StreamDescriptionSummary

Represents the output for DescribeStreamSummary

See: newStreamDescriptionSummary smart constructor.

Constructors

StreamDescriptionSummary'

Fields

encryptionType :: Maybe EncryptionType

The encryption type used. This value is one of the following:

KMS

NONE

keyId :: Maybe Text

The GUID for the customer-managed AWS KMS key to use for encryption. This value can be a globally unique identifier, a fully specified ARN to either an alias or a key, or an alias name prefixed by "alias/".You can also use a master key owned by Kinesis Data Streams by specifying the alias aws/kinesis.

Key ARN example: arn:aws:kms:us-east-1:123456789012:key/12345678-1234-1234-1234-123456789012

Alias ARN example: arn:aws:kms:us-east-1:123456789012:alias/MyAliasName

Globally unique key ID example: 12345678-1234-1234-1234-123456789012

Alias name example: alias/MyAliasName

Master key owned by Kinesis Data Streams: alias/aws/kinesis

consumerCount :: Maybe Natural

The number of enhanced fan-out consumers registered with the stream.

streamName :: Text

The name of the stream being described.

streamARN :: Text

The Amazon Resource Name (ARN) for the stream being described.

streamStatus :: StreamStatus

The current status of the stream being described. The stream status is one of the following states:

CREATING - The stream is being created. Kinesis Data Streams immediately returns and sets StreamStatus to CREATING.

DELETING - The stream is being deleted. The specified stream is in the DELETING state until Kinesis Data Streams completes the deletion.

ACTIVE - The stream exists and is ready for read and write operations or deletion. You should perform read and write operations only on an ACTIVE stream.

UPDATING - Shards in the stream are being merged or split. Read and write operations continue to work while the stream is in the UPDATING state.

retentionPeriodHours :: Int

The current retention period, in hours.

streamCreationTimestamp :: POSIX

The approximate time that the stream was created.

enhancedMonitoring :: [EnhancedMetrics]

Represents the current enhanced monitoring settings of the stream.

openShardCount :: Natural

The number of open shards in the stream.

newStreamDescriptionSummary

Arguments

:: Text
$sel:streamName:StreamDescriptionSummary'

-> Text
$sel:streamARN:StreamDescriptionSummary'

-> StreamStatus
$sel:streamStatus:StreamDescriptionSummary'

-> Int
$sel:retentionPeriodHours:StreamDescriptionSummary'

-> UTCTime
$sel:streamCreationTimestamp:StreamDescriptionSummary'

-> Natural
$sel:openShardCount:StreamDescriptionSummary'

-> StreamDescriptionSummary

Create a value of StreamDescriptionSummary with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:encryptionType:StreamDescriptionSummary', streamDescriptionSummary_encryptionType - The encryption type used. This value is one of the following:

KMS

NONE

$sel:keyId:StreamDescriptionSummary', streamDescriptionSummary_keyId - The GUID for the customer-managed AWS KMS key to use for encryption. This value can be a globally unique identifier, a fully specified ARN to either an alias or a key, or an alias name prefixed by "alias/".You can also use a master key owned by Kinesis Data Streams by specifying the alias aws/kinesis.

Key ARN example: arn:aws:kms:us-east-1:123456789012:key/12345678-1234-1234-1234-123456789012

Alias ARN example: arn:aws:kms:us-east-1:123456789012:alias/MyAliasName

Globally unique key ID example: 12345678-1234-1234-1234-123456789012

Alias name example: alias/MyAliasName

Master key owned by Kinesis Data Streams: alias/aws/kinesis

$sel:consumerCount:StreamDescriptionSummary', streamDescriptionSummary_consumerCount - The number of enhanced fan-out consumers registered with the stream.

$sel:streamName:StreamDescriptionSummary', streamDescriptionSummary_streamName - The name of the stream being described.

$sel:streamARN:StreamDescriptionSummary', streamDescriptionSummary_streamARN - The Amazon Resource Name (ARN) for the stream being described.

$sel:streamStatus:StreamDescriptionSummary', streamDescriptionSummary_streamStatus - The current status of the stream being described. The stream status is one of the following states:

CREATING - The stream is being created. Kinesis Data Streams immediately returns and sets StreamStatus to CREATING.

DELETING - The stream is being deleted. The specified stream is in the DELETING state until Kinesis Data Streams completes the deletion.

ACTIVE - The stream exists and is ready for read and write operations or deletion. You should perform read and write operations only on an ACTIVE stream.

UPDATING - Shards in the stream are being merged or split. Read and write operations continue to work while the stream is in the UPDATING state.

$sel:retentionPeriodHours:StreamDescriptionSummary', streamDescriptionSummary_retentionPeriodHours - The current retention period, in hours.

$sel:streamCreationTimestamp:StreamDescriptionSummary', streamDescriptionSummary_streamCreationTimestamp - The approximate time that the stream was created.

$sel:enhancedMonitoring:StreamDescriptionSummary', streamDescriptionSummary_enhancedMonitoring - Represents the current enhanced monitoring settings of the stream.

$sel:openShardCount:StreamDescriptionSummary', streamDescriptionSummary_openShardCount - The number of open shards in the stream.

streamDescriptionSummary_encryptionType :: Lens' StreamDescriptionSummary (Maybe EncryptionType)

The encryption type used. This value is one of the following:

KMS

NONE

streamDescriptionSummary_keyId :: Lens' StreamDescriptionSummary (Maybe Text)

The GUID for the customer-managed AWS KMS key to use for encryption. This value can be a globally unique identifier, a fully specified ARN to either an alias or a key, or an alias name prefixed by "alias/".You can also use a master key owned by Kinesis Data Streams by specifying the alias aws/kinesis.

Key ARN example: arn:aws:kms:us-east-1:123456789012:key/12345678-1234-1234-1234-123456789012

Alias ARN example: arn:aws:kms:us-east-1:123456789012:alias/MyAliasName

Globally unique key ID example: 12345678-1234-1234-1234-123456789012

Alias name example: alias/MyAliasName

Master key owned by Kinesis Data Streams: alias/aws/kinesis

streamDescriptionSummary_consumerCount :: Lens' StreamDescriptionSummary (Maybe Natural)

The number of enhanced fan-out consumers registered with the stream.

streamDescriptionSummary_streamName :: Lens' StreamDescriptionSummary Text

The name of the stream being described.

streamDescriptionSummary_streamARN :: Lens' StreamDescriptionSummary Text

The Amazon Resource Name (ARN) for the stream being described.

streamDescriptionSummary_streamStatus :: Lens' StreamDescriptionSummary StreamStatus

The current status of the stream being described. The stream status is one of the following states:

CREATING - The stream is being created. Kinesis Data Streams immediately returns and sets StreamStatus to CREATING.

DELETING - The stream is being deleted. The specified stream is in the DELETING state until Kinesis Data Streams completes the deletion.

ACTIVE - The stream exists and is ready for read and write operations or deletion. You should perform read and write operations only on an ACTIVE stream.

UPDATING - Shards in the stream are being merged or split. Read and write operations continue to work while the stream is in the UPDATING state.

streamDescriptionSummary_retentionPeriodHours :: Lens' StreamDescriptionSummary Int

The current retention period, in hours.

streamDescriptionSummary_streamCreationTimestamp :: Lens' StreamDescriptionSummary UTCTime

The approximate time that the stream was created.

streamDescriptionSummary_enhancedMonitoring :: Lens' StreamDescriptionSummary [EnhancedMetrics]

Represents the current enhanced monitoring settings of the stream.

streamDescriptionSummary_openShardCount :: Lens' StreamDescriptionSummary Natural

The number of open shards in the stream.