Amazonka.Kinesis.UpdateShardCount

Description

Updates the shard count of the specified stream to the specified number of shards.

Updating the shard count is an asynchronous operation. Upon receiving the request, Kinesis Data Streams returns immediately and sets the status of the stream to UPDATING. After the update is complete, Kinesis Data Streams sets the status of the stream back to ACTIVE. Depending on the size of the stream, the scaling action could take a few minutes to complete. You can continue to read and write data to your stream while its status is UPDATING.

To update the shard count, Kinesis Data Streams performs splits or merges on individual shards. This can cause short-lived shards to be created, in addition to the final shards. These short-lived shards count towards your total shard limit for your account in the Region.

When using this operation, we recommend that you specify a target shard count that is a multiple of 25% (25%, 50%, 75%, 100%). You can specify any target value within your shard limit. However, if you specify a target that isn't a multiple of 25%, the scaling action might take longer to complete.

This operation has the following default limits. By default, you cannot do the following:

Scale more than ten times per rolling 24-hour period per stream

Scale up to more than double your current shard count for a stream

Scale down below half your current shard count for a stream

Scale up to more than 500 shards in a stream

Scale a stream with more than 500 shards down unless the result is less than 500 shards

Scale up to more than the shard limit for your account

For the default limits for an AWS account, see Streams Limits in the Amazon Kinesis Data Streams Developer Guide. To request an increase in the call rate limit, the shard limit for this API, or your overall shard limit, use the limits form.

Creating a Request

data UpdateShardCount

See: newUpdateShardCount smart constructor.

Constructors

UpdateShardCount'

Fields

streamName :: Text

The name of the stream.

targetShardCount :: Natural

The new number of shards. This value has the following default limits. By default, you cannot do the following:

Set this value to more than double your current shard count for a stream.

Set this value below half your current shard count for a stream.

Set this value to more than 500 shards in a stream (the default limit for shard count per stream is 500 per account per region), unless you request a limit increase.

Scale a stream with more than 500 shards down unless you set this value to less than 500 shards.

scalingType :: ScalingType

The scaling type. Uniform scaling creates shards of equal size.

newUpdateShardCount

Arguments

:: Text
$sel:streamName:UpdateShardCount'

-> Natural
$sel:targetShardCount:UpdateShardCount'

-> ScalingType
$sel:scalingType:UpdateShardCount'

-> UpdateShardCount

Create a value of UpdateShardCount with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:streamName:UpdateShardCount', updateShardCount_streamName - The name of the stream.

$sel:targetShardCount:UpdateShardCount', updateShardCount_targetShardCount - The new number of shards. This value has the following default limits. By default, you cannot do the following:

Set this value to more than double your current shard count for a stream.

Set this value below half your current shard count for a stream.

Set this value to more than 500 shards in a stream (the default limit for shard count per stream is 500 per account per region), unless you request a limit increase.

Scale a stream with more than 500 shards down unless you set this value to less than 500 shards.

$sel:scalingType:UpdateShardCount', updateShardCount_scalingType - The scaling type. Uniform scaling creates shards of equal size.

Request Lenses

updateShardCount_streamName :: Lens' UpdateShardCount Text

The name of the stream.

updateShardCount_targetShardCount :: Lens' UpdateShardCount Natural

The new number of shards. This value has the following default limits. By default, you cannot do the following:

Set this value to more than double your current shard count for a stream.

Set this value below half your current shard count for a stream.

Set this value to more than 500 shards in a stream (the default limit for shard count per stream is 500 per account per region), unless you request a limit increase.

Scale a stream with more than 500 shards down unless you set this value to less than 500 shards.

updateShardCount_scalingType :: Lens' UpdateShardCount ScalingType

The scaling type. Uniform scaling creates shards of equal size.

Destructuring the Response

data UpdateShardCountResponse

See: newUpdateShardCountResponse smart constructor.

Constructors

UpdateShardCountResponse'

Fields

targetShardCount :: Maybe Natural

The updated number of shards.

streamName :: Maybe Text

The name of the stream.

currentShardCount :: Maybe Natural

The current number of shards.

httpStatus :: Int

The response's http status code.

newUpdateShardCountResponse

Arguments

:: Int
$sel:httpStatus:UpdateShardCountResponse'

-> UpdateShardCountResponse

Create a value of UpdateShardCountResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:targetShardCount:UpdateShardCount', updateShardCountResponse_targetShardCount - The updated number of shards.

$sel:streamName:UpdateShardCount', updateShardCountResponse_streamName - The name of the stream.

$sel:currentShardCount:UpdateShardCountResponse', updateShardCountResponse_currentShardCount - The current number of shards.

$sel:httpStatus:UpdateShardCountResponse', updateShardCountResponse_httpStatus - The response's http status code.

Response Lenses

updateShardCountResponse_targetShardCount :: Lens' UpdateShardCountResponse (Maybe Natural)

The updated number of shards.

updateShardCountResponse_streamName :: Lens' UpdateShardCountResponse (Maybe Text)

The name of the stream.

updateShardCountResponse_currentShardCount :: Lens' UpdateShardCountResponse (Maybe Natural)

The current number of shards.

updateShardCountResponse_httpStatus :: Lens' UpdateShardCountResponse Int

The response's http status code.