Amazonka.KinesisAnalytics.Types.CSVMappingParameters

Description

Documentation

data CSVMappingParameters

Provides additional mapping information when the record format uses delimiters, such as CSV. For example, the following sample records use CSV format, where the records use the '\n' as the row delimiter and a comma (",") as the column delimiter:

"name1", "address1"

"name2", "address2"

See: newCSVMappingParameters smart constructor.

Constructors

CSVMappingParameters'

Fields

recordRowDelimiter :: Text

Row delimiter. For example, in a CSV format, '\n' is the typical row delimiter.

recordColumnDelimiter :: Text

Column delimiter. For example, in a CSV format, a comma (",") is the typical column delimiter.

newCSVMappingParameters

Arguments

:: Text
$sel:recordRowDelimiter:CSVMappingParameters'

-> Text
$sel:recordColumnDelimiter:CSVMappingParameters'

-> CSVMappingParameters

Create a value of CSVMappingParameters with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:recordRowDelimiter:CSVMappingParameters', cSVMappingParameters_recordRowDelimiter - Row delimiter. For example, in a CSV format, '\n' is the typical row delimiter.

$sel:recordColumnDelimiter:CSVMappingParameters', cSVMappingParameters_recordColumnDelimiter - Column delimiter. For example, in a CSV format, a comma (",") is the typical column delimiter.

cSVMappingParameters_recordRowDelimiter :: Lens' CSVMappingParameters Text

Row delimiter. For example, in a CSV format, '\n' is the typical row delimiter.

cSVMappingParameters_recordColumnDelimiter :: Lens' CSVMappingParameters Text

Column delimiter. For example, in a CSV format, a comma (",") is the typical column delimiter.