Amazonka.KinesisAnalytics.Types.KinesisStreamsOutputUpdate

Description

Documentation

data KinesisStreamsOutputUpdate

When updating an output configuration using the UpdateApplication operation, provides information about an Amazon Kinesis stream configured as the destination.

See: newKinesisStreamsOutputUpdate smart constructor.

Constructors

KinesisStreamsOutputUpdate'

Fields

roleARNUpdate :: Maybe Text

ARN of the IAM role that Amazon Kinesis Analytics can assume to access the stream on your behalf. You need to grant the necessary permissions to this role.

resourceARNUpdate :: Maybe Text

Amazon Resource Name (ARN) of the Amazon Kinesis stream where you want to write the output.

newKinesisStreamsOutputUpdate :: KinesisStreamsOutputUpdate

Create a value of KinesisStreamsOutputUpdate with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:roleARNUpdate:KinesisStreamsOutputUpdate', kinesisStreamsOutputUpdate_roleARNUpdate - ARN of the IAM role that Amazon Kinesis Analytics can assume to access the stream on your behalf. You need to grant the necessary permissions to this role.

$sel:resourceARNUpdate:KinesisStreamsOutputUpdate', kinesisStreamsOutputUpdate_resourceARNUpdate - Amazon Resource Name (ARN) of the Amazon Kinesis stream where you want to write the output.

kinesisStreamsOutputUpdate_roleARNUpdate :: Lens' KinesisStreamsOutputUpdate (Maybe Text)

ARN of the IAM role that Amazon Kinesis Analytics can assume to access the stream on your behalf. You need to grant the necessary permissions to this role.

kinesisStreamsOutputUpdate_resourceARNUpdate :: Lens' KinesisStreamsOutputUpdate (Maybe Text)

Amazon Resource Name (ARN) of the Amazon Kinesis stream where you want to write the output.