Amazonka.KinesisAnalytics.Types.OutputUpdate

Description

Documentation

data OutputUpdate

Describes updates to the output configuration identified by the OutputId.

See: newOutputUpdate smart constructor.

Constructors

OutputUpdate'

Fields

kinesisStreamsOutputUpdate :: Maybe KinesisStreamsOutputUpdate

Describes an Amazon Kinesis stream as the destination for the output.

destinationSchemaUpdate :: Maybe DestinationSchema

Describes the data format when records are written to the destination. For more information, see Configuring Application Output.

kinesisFirehoseOutputUpdate :: Maybe KinesisFirehoseOutputUpdate

Describes an Amazon Kinesis Firehose delivery stream as the destination for the output.

nameUpdate :: Maybe Text

If you want to specify a different in-application stream for this output configuration, use this field to specify the new in-application stream name.

lambdaOutputUpdate :: Maybe LambdaOutputUpdate

Describes an AWS Lambda function as the destination for the output.

outputId :: Text

Identifies the specific output configuration that you want to update.

newOutputUpdate

Arguments

:: Text
$sel:outputId:OutputUpdate'

-> OutputUpdate

Create a value of OutputUpdate with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:kinesisStreamsOutputUpdate:OutputUpdate', outputUpdate_kinesisStreamsOutputUpdate - Describes an Amazon Kinesis stream as the destination for the output.

$sel:destinationSchemaUpdate:OutputUpdate', outputUpdate_destinationSchemaUpdate - Describes the data format when records are written to the destination. For more information, see Configuring Application Output.

$sel:kinesisFirehoseOutputUpdate:OutputUpdate', outputUpdate_kinesisFirehoseOutputUpdate - Describes an Amazon Kinesis Firehose delivery stream as the destination for the output.

$sel:nameUpdate:OutputUpdate', outputUpdate_nameUpdate - If you want to specify a different in-application stream for this output configuration, use this field to specify the new in-application stream name.

$sel:lambdaOutputUpdate:OutputUpdate', outputUpdate_lambdaOutputUpdate - Describes an AWS Lambda function as the destination for the output.

$sel:outputId:OutputUpdate', outputUpdate_outputId - Identifies the specific output configuration that you want to update.

outputUpdate_kinesisStreamsOutputUpdate :: Lens' OutputUpdate (Maybe KinesisStreamsOutputUpdate)

Describes an Amazon Kinesis stream as the destination for the output.

outputUpdate_destinationSchemaUpdate :: Lens' OutputUpdate (Maybe DestinationSchema)

Describes the data format when records are written to the destination. For more information, see Configuring Application Output.

outputUpdate_kinesisFirehoseOutputUpdate :: Lens' OutputUpdate (Maybe KinesisFirehoseOutputUpdate)

Describes an Amazon Kinesis Firehose delivery stream as the destination for the output.

outputUpdate_nameUpdate :: Lens' OutputUpdate (Maybe Text)

If you want to specify a different in-application stream for this output configuration, use this field to specify the new in-application stream name.

outputUpdate_lambdaOutputUpdate :: Lens' OutputUpdate (Maybe LambdaOutputUpdate)

Describes an AWS Lambda function as the destination for the output.

outputUpdate_outputId :: Lens' OutputUpdate Text

Identifies the specific output configuration that you want to update.