Amazonka.KinesisAnalyticsV2.AddApplicationOutput

Description

Adds an external destination to your SQL-based Kinesis Data Analytics application.

If you want Kinesis Data Analytics to deliver data from an in-application stream within your application to an external destination (such as an Kinesis data stream, a Kinesis Data Firehose delivery stream, or an Amazon Lambda function), you add the relevant configuration to your application using this operation. You can configure one or more outputs for your application. Each output configuration maps an in-application stream and an external destination.

You can use one of the output configurations to deliver data from your in-application error stream to an external destination so that you can analyze the errors.

Any configuration update, including adding a streaming source using this operation, results in a new version of the application. You can use the DescribeApplication operation to find the current application version.

Creating a Request

data AddApplicationOutput

See: newAddApplicationOutput smart constructor.

Constructors

AddApplicationOutput'

Fields

applicationName :: Text

The name of the application to which you want to add the output configuration.

currentApplicationVersionId :: Natural

The version of the application to which you want to add the output configuration. You can use the DescribeApplication operation to get the current application version. If the version specified is not the current version, the ConcurrentModificationException is returned.

output :: Output

An array of objects, each describing one output configuration. In the output configuration, you specify the name of an in-application stream, a destination (that is, a Kinesis data stream, a Kinesis Data Firehose delivery stream, or an Amazon Lambda function), and record the formation to use when writing to the destination.

newAddApplicationOutput

Arguments

:: Text
$sel:applicationName:AddApplicationOutput'

-> Natural
$sel:currentApplicationVersionId:AddApplicationOutput'

-> Output
$sel:output:AddApplicationOutput'

-> AddApplicationOutput

Create a value of AddApplicationOutput with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:applicationName:AddApplicationOutput', addApplicationOutput_applicationName - The name of the application to which you want to add the output configuration.

$sel:currentApplicationVersionId:AddApplicationOutput', addApplicationOutput_currentApplicationVersionId - The version of the application to which you want to add the output configuration. You can use the DescribeApplication operation to get the current application version. If the version specified is not the current version, the ConcurrentModificationException is returned.

$sel:output:AddApplicationOutput', addApplicationOutput_output - An array of objects, each describing one output configuration. In the output configuration, you specify the name of an in-application stream, a destination (that is, a Kinesis data stream, a Kinesis Data Firehose delivery stream, or an Amazon Lambda function), and record the formation to use when writing to the destination.

Request Lenses

addApplicationOutput_applicationName :: Lens' AddApplicationOutput Text

The name of the application to which you want to add the output configuration.

addApplicationOutput_currentApplicationVersionId :: Lens' AddApplicationOutput Natural

The version of the application to which you want to add the output configuration. You can use the DescribeApplication operation to get the current application version. If the version specified is not the current version, the ConcurrentModificationException is returned.

addApplicationOutput_output :: Lens' AddApplicationOutput Output

An array of objects, each describing one output configuration. In the output configuration, you specify the name of an in-application stream, a destination (that is, a Kinesis data stream, a Kinesis Data Firehose delivery stream, or an Amazon Lambda function), and record the formation to use when writing to the destination.

Destructuring the Response

data AddApplicationOutputResponse

See: newAddApplicationOutputResponse smart constructor.

Constructors

AddApplicationOutputResponse'

Fields

applicationARN :: Maybe Text

The application Amazon Resource Name (ARN).

outputDescriptions :: Maybe [OutputDescription]

Describes the application output configuration. For more information, see Configuring Application Output.

applicationVersionId :: Maybe Natural

The updated application version ID. Kinesis Data Analytics increments this ID when the application is updated.

httpStatus :: Int

The response's http status code.

newAddApplicationOutputResponse

Arguments

:: Int
$sel:httpStatus:AddApplicationOutputResponse'

-> AddApplicationOutputResponse

Create a value of AddApplicationOutputResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:applicationARN:AddApplicationOutputResponse', addApplicationOutputResponse_applicationARN - The application Amazon Resource Name (ARN).

$sel:outputDescriptions:AddApplicationOutputResponse', addApplicationOutputResponse_outputDescriptions - Describes the application output configuration. For more information, see Configuring Application Output.

$sel:applicationVersionId:AddApplicationOutputResponse', addApplicationOutputResponse_applicationVersionId - The updated application version ID. Kinesis Data Analytics increments this ID when the application is updated.

$sel:httpStatus:AddApplicationOutputResponse', addApplicationOutputResponse_httpStatus - The response's http status code.

Response Lenses

addApplicationOutputResponse_applicationARN :: Lens' AddApplicationOutputResponse (Maybe Text)

The application Amazon Resource Name (ARN).

addApplicationOutputResponse_outputDescriptions :: Lens' AddApplicationOutputResponse (Maybe [OutputDescription])

Describes the application output configuration. For more information, see Configuring Application Output.

addApplicationOutputResponse_applicationVersionId :: Lens' AddApplicationOutputResponse (Maybe Natural)

The updated application version ID. Kinesis Data Analytics increments this ID when the application is updated.

addApplicationOutputResponse_httpStatus :: Lens' AddApplicationOutputResponse Int

The response's http status code.