Amazonka.KinesisAnalyticsV2.DeleteApplicationOutput

Description

Deletes the output destination configuration from your SQL-based Kinesis Data Analytics application's configuration. Kinesis Data Analytics will no longer write data from the corresponding in-application stream to the external output destination.

Creating a Request

data DeleteApplicationOutput

See: newDeleteApplicationOutput smart constructor.

Constructors

DeleteApplicationOutput'

Fields

applicationName :: Text

The application name.

currentApplicationVersionId :: Natural

The application version. You can use the DescribeApplication operation to get the current application version. If the version specified is not the current version, the ConcurrentModificationException is returned.

outputId :: Text

The ID of the configuration to delete. Each output configuration that is added to the application (either when the application is created or later) using the AddApplicationOutput operation has a unique ID. You need to provide the ID to uniquely identify the output configuration that you want to delete from the application configuration. You can use the DescribeApplication operation to get the specific OutputId.

newDeleteApplicationOutput

Arguments

:: Text
$sel:applicationName:DeleteApplicationOutput'

-> Natural
$sel:currentApplicationVersionId:DeleteApplicationOutput'

-> Text
$sel:outputId:DeleteApplicationOutput'

-> DeleteApplicationOutput

Create a value of DeleteApplicationOutput with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:applicationName:DeleteApplicationOutput', deleteApplicationOutput_applicationName - The application name.

$sel:currentApplicationVersionId:DeleteApplicationOutput', deleteApplicationOutput_currentApplicationVersionId - The application version. You can use the DescribeApplication operation to get the current application version. If the version specified is not the current version, the ConcurrentModificationException is returned.

$sel:outputId:DeleteApplicationOutput', deleteApplicationOutput_outputId - The ID of the configuration to delete. Each output configuration that is added to the application (either when the application is created or later) using the AddApplicationOutput operation has a unique ID. You need to provide the ID to uniquely identify the output configuration that you want to delete from the application configuration. You can use the DescribeApplication operation to get the specific OutputId.

Request Lenses

deleteApplicationOutput_applicationName :: Lens' DeleteApplicationOutput Text

The application name.

deleteApplicationOutput_currentApplicationVersionId :: Lens' DeleteApplicationOutput Natural

The application version. You can use the DescribeApplication operation to get the current application version. If the version specified is not the current version, the ConcurrentModificationException is returned.

deleteApplicationOutput_outputId :: Lens' DeleteApplicationOutput Text

The ID of the configuration to delete. Each output configuration that is added to the application (either when the application is created or later) using the AddApplicationOutput operation has a unique ID. You need to provide the ID to uniquely identify the output configuration that you want to delete from the application configuration. You can use the DescribeApplication operation to get the specific OutputId.

Destructuring the Response

data DeleteApplicationOutputResponse

See: newDeleteApplicationOutputResponse smart constructor.

Constructors

DeleteApplicationOutputResponse'

Fields

applicationARN :: Maybe Text

The application Amazon Resource Name (ARN).

applicationVersionId :: Maybe Natural

The current application version ID.

httpStatus :: Int

The response's http status code.

newDeleteApplicationOutputResponse

Arguments

:: Int
$sel:httpStatus:DeleteApplicationOutputResponse'

-> DeleteApplicationOutputResponse

Create a value of DeleteApplicationOutputResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:applicationARN:DeleteApplicationOutputResponse', deleteApplicationOutputResponse_applicationARN - The application Amazon Resource Name (ARN).

$sel:applicationVersionId:DeleteApplicationOutputResponse', deleteApplicationOutputResponse_applicationVersionId - The current application version ID.

$sel:httpStatus:DeleteApplicationOutputResponse', deleteApplicationOutputResponse_httpStatus - The response's http status code.

Response Lenses

deleteApplicationOutputResponse_applicationARN :: Lens' DeleteApplicationOutputResponse (Maybe Text)

The application Amazon Resource Name (ARN).

deleteApplicationOutputResponse_applicationVersionId :: Lens' DeleteApplicationOutputResponse (Maybe Natural)

The current application version ID.

deleteApplicationOutputResponse_httpStatus :: Lens' DeleteApplicationOutputResponse Int

The response's http status code.