Amazonka.KinesisAnalyticsV2.ListApplicationVersions

Description

Lists all the versions for the specified application, including versions that were rolled back. The response also includes a summary of the configuration associated with each version.

To get the complete description of a specific application version, invoke the DescribeApplicationVersion operation.

This operation is supported only for Amazon Kinesis Data Analytics for Apache Flink.

Creating a Request

data ListApplicationVersions

See: newListApplicationVersions smart constructor.

Constructors

ListApplicationVersions'

Fields

nextToken :: Maybe Text

If a previous invocation of this operation returned a pagination token, pass it into this value to retrieve the next set of results. For more information about pagination, see Using the Amazon Command Line Interface's Pagination Options.

limit :: Maybe Natural

The maximum number of versions to list in this invocation of the operation.

applicationName :: Text

The name of the application for which you want to list all versions.

newListApplicationVersions

Arguments

:: Text
$sel:applicationName:ListApplicationVersions'

-> ListApplicationVersions

Create a value of ListApplicationVersions with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:nextToken:ListApplicationVersions', listApplicationVersions_nextToken - If a previous invocation of this operation returned a pagination token, pass it into this value to retrieve the next set of results. For more information about pagination, see Using the Amazon Command Line Interface's Pagination Options.

$sel:limit:ListApplicationVersions', listApplicationVersions_limit - The maximum number of versions to list in this invocation of the operation.

$sel:applicationName:ListApplicationVersions', listApplicationVersions_applicationName - The name of the application for which you want to list all versions.

Request Lenses

listApplicationVersions_nextToken :: Lens' ListApplicationVersions (Maybe Text)

If a previous invocation of this operation returned a pagination token, pass it into this value to retrieve the next set of results. For more information about pagination, see Using the Amazon Command Line Interface's Pagination Options.

listApplicationVersions_limit :: Lens' ListApplicationVersions (Maybe Natural)

The maximum number of versions to list in this invocation of the operation.

listApplicationVersions_applicationName :: Lens' ListApplicationVersions Text

The name of the application for which you want to list all versions.

Destructuring the Response

data ListApplicationVersionsResponse

See: newListApplicationVersionsResponse smart constructor.

Constructors

ListApplicationVersionsResponse'

Fields

applicationVersionSummaries :: Maybe [ApplicationVersionSummary]

A list of the application versions and the associated configuration summaries. The list includes application versions that were rolled back.

To get the complete description of a specific application version, invoke the DescribeApplicationVersion operation.

nextToken :: Maybe Text

The pagination token for the next set of results, or null if there are no additional results. To retrieve the next set of items, pass this token into a subsequent invocation of this operation. For more information about pagination, see Using the Amazon Command Line Interface's Pagination Options.

httpStatus :: Int

The response's http status code.

newListApplicationVersionsResponse

Arguments

:: Int
$sel:httpStatus:ListApplicationVersionsResponse'

-> ListApplicationVersionsResponse

Create a value of ListApplicationVersionsResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:applicationVersionSummaries:ListApplicationVersionsResponse', listApplicationVersionsResponse_applicationVersionSummaries - A list of the application versions and the associated configuration summaries. The list includes application versions that were rolled back.

To get the complete description of a specific application version, invoke the DescribeApplicationVersion operation.

$sel:nextToken:ListApplicationVersions', listApplicationVersionsResponse_nextToken - The pagination token for the next set of results, or null if there are no additional results. To retrieve the next set of items, pass this token into a subsequent invocation of this operation. For more information about pagination, see Using the Amazon Command Line Interface's Pagination Options.

$sel:httpStatus:ListApplicationVersionsResponse', listApplicationVersionsResponse_httpStatus - The response's http status code.

Response Lenses

listApplicationVersionsResponse_applicationVersionSummaries :: Lens' ListApplicationVersionsResponse (Maybe [ApplicationVersionSummary])

A list of the application versions and the associated configuration summaries. The list includes application versions that were rolled back.

To get the complete description of a specific application version, invoke the DescribeApplicationVersion operation.

listApplicationVersionsResponse_nextToken :: Lens' ListApplicationVersionsResponse (Maybe Text)

The pagination token for the next set of results, or null if there are no additional results. To retrieve the next set of items, pass this token into a subsequent invocation of this operation. For more information about pagination, see Using the Amazon Command Line Interface's Pagination Options.

listApplicationVersionsResponse_httpStatus :: Lens' ListApplicationVersionsResponse Int

The response's http status code.