Amazonka.KinesisAnalyticsV2.Types.ApplicationCodeConfiguration

Description

Documentation

data ApplicationCodeConfiguration

Describes code configuration for an application.

See: newApplicationCodeConfiguration smart constructor.

Constructors

ApplicationCodeConfiguration'

Fields

codeContent :: Maybe CodeContent

The location and type of the application code.

codeContentType :: CodeContentType

Specifies whether the code content is in text or zip format.

newApplicationCodeConfiguration

Arguments

:: CodeContentType
$sel:codeContentType:ApplicationCodeConfiguration'

-> ApplicationCodeConfiguration

Create a value of ApplicationCodeConfiguration with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:codeContent:ApplicationCodeConfiguration', applicationCodeConfiguration_codeContent - The location and type of the application code.

$sel:codeContentType:ApplicationCodeConfiguration', applicationCodeConfiguration_codeContentType - Specifies whether the code content is in text or zip format.

applicationCodeConfiguration_codeContent :: Lens' ApplicationCodeConfiguration (Maybe CodeContent)

The location and type of the application code.

applicationCodeConfiguration_codeContentType :: Lens' ApplicationCodeConfiguration CodeContentType

Specifies whether the code content is in text or zip format.