Amazonka.KinesisAnalyticsV2.Types.ApplicationCodeConfigurationUpdate

Description

Documentation

data ApplicationCodeConfigurationUpdate

Describes code configuration updates for an application. This is supported for a Flink-based Kinesis Data Analytics application or a SQL-based Kinesis Data Analytics application.

See: newApplicationCodeConfigurationUpdate smart constructor.

Constructors

ApplicationCodeConfigurationUpdate'

Fields

codeContentTypeUpdate :: Maybe CodeContentType

Describes updates to the code content type.

codeContentUpdate :: Maybe CodeContentUpdate

Describes updates to the code content of an application.

newApplicationCodeConfigurationUpdate :: ApplicationCodeConfigurationUpdate

Create a value of ApplicationCodeConfigurationUpdate with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:codeContentTypeUpdate:ApplicationCodeConfigurationUpdate', applicationCodeConfigurationUpdate_codeContentTypeUpdate - Describes updates to the code content type.

$sel:codeContentUpdate:ApplicationCodeConfigurationUpdate', applicationCodeConfigurationUpdate_codeContentUpdate - Describes updates to the code content of an application.

applicationCodeConfigurationUpdate_codeContentTypeUpdate :: Lens' ApplicationCodeConfigurationUpdate (Maybe CodeContentType)

Describes updates to the code content type.

applicationCodeConfigurationUpdate_codeContentUpdate :: Lens' ApplicationCodeConfigurationUpdate (Maybe CodeContentUpdate)

Describes updates to the code content of an application.