Amazonka.KinesisAnalyticsV2.Types.ApplicationConfigurationUpdate

Description

Documentation

data ApplicationConfigurationUpdate

Describes updates to an application's configuration.

See: newApplicationConfigurationUpdate smart constructor.

Constructors

ApplicationConfigurationUpdate'

Fields

applicationSnapshotConfigurationUpdate :: Maybe ApplicationSnapshotConfigurationUpdate

Describes whether snapshots are enabled for a Flink-based Kinesis Data Analytics application.

zeppelinApplicationConfigurationUpdate :: Maybe ZeppelinApplicationConfigurationUpdate

Updates to the configuration of a Kinesis Data Analytics Studio notebook.

flinkApplicationConfigurationUpdate :: Maybe FlinkApplicationConfigurationUpdate

Describes updates to a Flink-based Kinesis Data Analytics application's configuration.

sqlApplicationConfigurationUpdate :: Maybe SqlApplicationConfigurationUpdate

Describes updates to a SQL-based Kinesis Data Analytics application's configuration.

vpcConfigurationUpdates :: Maybe [VpcConfigurationUpdate]

Updates to the array of descriptions of VPC configurations available to the application.

environmentPropertyUpdates :: Maybe EnvironmentPropertyUpdates

Describes updates to the environment properties for a Flink-based Kinesis Data Analytics application.

applicationCodeConfigurationUpdate :: Maybe ApplicationCodeConfigurationUpdate

Describes updates to an application's code configuration.

newApplicationConfigurationUpdate :: ApplicationConfigurationUpdate

Create a value of ApplicationConfigurationUpdate with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:applicationSnapshotConfigurationUpdate:ApplicationConfigurationUpdate', applicationConfigurationUpdate_applicationSnapshotConfigurationUpdate - Describes whether snapshots are enabled for a Flink-based Kinesis Data Analytics application.

$sel:zeppelinApplicationConfigurationUpdate:ApplicationConfigurationUpdate', applicationConfigurationUpdate_zeppelinApplicationConfigurationUpdate - Updates to the configuration of a Kinesis Data Analytics Studio notebook.

$sel:flinkApplicationConfigurationUpdate:ApplicationConfigurationUpdate', applicationConfigurationUpdate_flinkApplicationConfigurationUpdate - Describes updates to a Flink-based Kinesis Data Analytics application's configuration.

$sel:sqlApplicationConfigurationUpdate:ApplicationConfigurationUpdate', applicationConfigurationUpdate_sqlApplicationConfigurationUpdate - Describes updates to a SQL-based Kinesis Data Analytics application's configuration.

$sel:vpcConfigurationUpdates:ApplicationConfigurationUpdate', applicationConfigurationUpdate_vpcConfigurationUpdates - Updates to the array of descriptions of VPC configurations available to the application.

$sel:environmentPropertyUpdates:ApplicationConfigurationUpdate', applicationConfigurationUpdate_environmentPropertyUpdates - Describes updates to the environment properties for a Flink-based Kinesis Data Analytics application.

$sel:applicationCodeConfigurationUpdate:ApplicationConfigurationUpdate', applicationConfigurationUpdate_applicationCodeConfigurationUpdate - Describes updates to an application's code configuration.

applicationConfigurationUpdate_applicationSnapshotConfigurationUpdate :: Lens' ApplicationConfigurationUpdate (Maybe ApplicationSnapshotConfigurationUpdate)

Describes whether snapshots are enabled for a Flink-based Kinesis Data Analytics application.

applicationConfigurationUpdate_zeppelinApplicationConfigurationUpdate :: Lens' ApplicationConfigurationUpdate (Maybe ZeppelinApplicationConfigurationUpdate)

Updates to the configuration of a Kinesis Data Analytics Studio notebook.

applicationConfigurationUpdate_flinkApplicationConfigurationUpdate :: Lens' ApplicationConfigurationUpdate (Maybe FlinkApplicationConfigurationUpdate)

Describes updates to a Flink-based Kinesis Data Analytics application's configuration.

applicationConfigurationUpdate_sqlApplicationConfigurationUpdate :: Lens' ApplicationConfigurationUpdate (Maybe SqlApplicationConfigurationUpdate)

Describes updates to a SQL-based Kinesis Data Analytics application's configuration.

applicationConfigurationUpdate_vpcConfigurationUpdates :: Lens' ApplicationConfigurationUpdate (Maybe [VpcConfigurationUpdate])

Updates to the array of descriptions of VPC configurations available to the application.

applicationConfigurationUpdate_environmentPropertyUpdates :: Lens' ApplicationConfigurationUpdate (Maybe EnvironmentPropertyUpdates)

Describes updates to the environment properties for a Flink-based Kinesis Data Analytics application.

applicationConfigurationUpdate_applicationCodeConfigurationUpdate :: Lens' ApplicationConfigurationUpdate (Maybe ApplicationCodeConfigurationUpdate)

Describes updates to an application's code configuration.