Amazonka.KinesisAnalyticsV2.Types.CustomArtifactConfigurationDescription

Description

Documentation

data CustomArtifactConfigurationDescription

Specifies a dependency JAR or a JAR of user-defined functions.

See: newCustomArtifactConfigurationDescription smart constructor.

Constructors

CustomArtifactConfigurationDescription'

Fields

mavenReferenceDescription :: Maybe MavenReference

The parameters that are required to specify a Maven dependency.

s3ContentLocationDescription :: Maybe S3ContentLocation

artifactType :: Maybe ArtifactType

UDF stands for user-defined functions. This type of artifact must be in an S3 bucket. A DEPENDENCY_JAR can be in either Maven or an S3 bucket.

newCustomArtifactConfigurationDescription :: CustomArtifactConfigurationDescription

Create a value of CustomArtifactConfigurationDescription with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:mavenReferenceDescription:CustomArtifactConfigurationDescription', customArtifactConfigurationDescription_mavenReferenceDescription - The parameters that are required to specify a Maven dependency.

$sel:s3ContentLocationDescription:CustomArtifactConfigurationDescription', customArtifactConfigurationDescription_s3ContentLocationDescription - Undocumented member.

$sel:artifactType:CustomArtifactConfigurationDescription', customArtifactConfigurationDescription_artifactType - UDF stands for user-defined functions. This type of artifact must be in an S3 bucket. A DEPENDENCY_JAR can be in either Maven or an S3 bucket.

customArtifactConfigurationDescription_mavenReferenceDescription :: Lens' CustomArtifactConfigurationDescription (Maybe MavenReference)

The parameters that are required to specify a Maven dependency.

customArtifactConfigurationDescription_s3ContentLocationDescription :: Lens' CustomArtifactConfigurationDescription (Maybe S3ContentLocation)

Undocumented member.

customArtifactConfigurationDescription_artifactType :: Lens' CustomArtifactConfigurationDescription (Maybe ArtifactType)

UDF stands for user-defined functions. This type of artifact must be in an S3 bucket. A DEPENDENCY_JAR can be in either Maven or an S3 bucket.