Amazonka.KinesisAnalyticsV2.Types.FlinkApplicationConfigurationDescription

Description

Documentation

data FlinkApplicationConfigurationDescription

Describes configuration parameters for a Flink-based Kinesis Data Analytics application.

See: newFlinkApplicationConfigurationDescription smart constructor.

Constructors

FlinkApplicationConfigurationDescription'

Fields

monitoringConfigurationDescription :: Maybe MonitoringConfigurationDescription

Describes configuration parameters for Amazon CloudWatch logging for an application.

parallelismConfigurationDescription :: Maybe ParallelismConfigurationDescription

Describes parameters for how an application executes multiple tasks simultaneously.

checkpointConfigurationDescription :: Maybe CheckpointConfigurationDescription

Describes an application's checkpointing configuration. Checkpointing is the process of persisting application state for fault tolerance.

jobPlanDescription :: Maybe Text

The job plan for an application. For more information about the job plan, see Jobs and Scheduling in the Apache Flink Documentation. To retrieve the job plan for the application, use the DescribeApplicationRequest$IncludeAdditionalDetails parameter of the DescribeApplication operation.

newFlinkApplicationConfigurationDescription :: FlinkApplicationConfigurationDescription

Create a value of FlinkApplicationConfigurationDescription with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:monitoringConfigurationDescription:FlinkApplicationConfigurationDescription', flinkApplicationConfigurationDescription_monitoringConfigurationDescription - Describes configuration parameters for Amazon CloudWatch logging for an application.

$sel:parallelismConfigurationDescription:FlinkApplicationConfigurationDescription', flinkApplicationConfigurationDescription_parallelismConfigurationDescription - Describes parameters for how an application executes multiple tasks simultaneously.

$sel:checkpointConfigurationDescription:FlinkApplicationConfigurationDescription', flinkApplicationConfigurationDescription_checkpointConfigurationDescription - Describes an application's checkpointing configuration. Checkpointing is the process of persisting application state for fault tolerance.

$sel:jobPlanDescription:FlinkApplicationConfigurationDescription', flinkApplicationConfigurationDescription_jobPlanDescription - The job plan for an application. For more information about the job plan, see Jobs and Scheduling in the Apache Flink Documentation. To retrieve the job plan for the application, use the DescribeApplicationRequest$IncludeAdditionalDetails parameter of the DescribeApplication operation.

flinkApplicationConfigurationDescription_monitoringConfigurationDescription :: Lens' FlinkApplicationConfigurationDescription (Maybe MonitoringConfigurationDescription)

Describes configuration parameters for Amazon CloudWatch logging for an application.

flinkApplicationConfigurationDescription_parallelismConfigurationDescription :: Lens' FlinkApplicationConfigurationDescription (Maybe ParallelismConfigurationDescription)

Describes parameters for how an application executes multiple tasks simultaneously.

flinkApplicationConfigurationDescription_checkpointConfigurationDescription :: Lens' FlinkApplicationConfigurationDescription (Maybe CheckpointConfigurationDescription)

Describes an application's checkpointing configuration. Checkpointing is the process of persisting application state for fault tolerance.

flinkApplicationConfigurationDescription_jobPlanDescription :: Lens' FlinkApplicationConfigurationDescription (Maybe Text)

The job plan for an application. For more information about the job plan, see Jobs and Scheduling in the Apache Flink Documentation. To retrieve the job plan for the application, use the DescribeApplicationRequest$IncludeAdditionalDetails parameter of the DescribeApplication operation.