Amazonka.KinesisAnalyticsV2.Types.MappingParameters

Description

Documentation

data MappingParameters

When you configure a SQL-based Kinesis Data Analytics application's input at the time of creating or updating an application, provides additional mapping information specific to the record format (such as JSON, CSV, or record fields delimited by some delimiter) on the streaming source.

See: newMappingParameters smart constructor.

Constructors

MappingParameters'

Fields

cSVMappingParameters :: Maybe CSVMappingParameters

Provides additional mapping information when the record format uses delimiters (for example, CSV).

jSONMappingParameters :: Maybe JSONMappingParameters

Provides additional mapping information when JSON is the record format on the streaming source.

newMappingParameters :: MappingParameters

Create a value of MappingParameters with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:cSVMappingParameters:MappingParameters', mappingParameters_cSVMappingParameters - Provides additional mapping information when the record format uses delimiters (for example, CSV).

$sel:jSONMappingParameters:MappingParameters', mappingParameters_jSONMappingParameters - Provides additional mapping information when JSON is the record format on the streaming source.

mappingParameters_cSVMappingParameters :: Lens' MappingParameters (Maybe CSVMappingParameters)

Provides additional mapping information when the record format uses delimiters (for example, CSV).

mappingParameters_jSONMappingParameters :: Lens' MappingParameters (Maybe JSONMappingParameters)

Provides additional mapping information when JSON is the record format on the streaming source.