Amazonka.KinesisAnalyticsV2.Types.MonitoringConfiguration

Description

Documentation

data MonitoringConfiguration

Describes configuration parameters for Amazon CloudWatch logging for an application. For more information about CloudWatch logging, see Monitoring.

See: newMonitoringConfiguration smart constructor.

Constructors

MonitoringConfiguration'

Fields

logLevel :: Maybe LogLevel

Describes the verbosity of the CloudWatch Logs for an application.

metricsLevel :: Maybe MetricsLevel

Describes the granularity of the CloudWatch Logs for an application. The Parallelism level is not recommended for applications with a Parallelism over 64 due to excessive costs.

configurationType :: ConfigurationType

Describes whether to use the default CloudWatch logging configuration for an application. You must set this property to CUSTOM in order to set the LogLevel or MetricsLevel parameters.

newMonitoringConfiguration

Arguments

:: ConfigurationType
$sel:configurationType:MonitoringConfiguration'

-> MonitoringConfiguration

Create a value of MonitoringConfiguration with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:logLevel:MonitoringConfiguration', monitoringConfiguration_logLevel - Describes the verbosity of the CloudWatch Logs for an application.

$sel:metricsLevel:MonitoringConfiguration', monitoringConfiguration_metricsLevel - Describes the granularity of the CloudWatch Logs for an application. The Parallelism level is not recommended for applications with a Parallelism over 64 due to excessive costs.

$sel:configurationType:MonitoringConfiguration', monitoringConfiguration_configurationType - Describes whether to use the default CloudWatch logging configuration for an application. You must set this property to CUSTOM in order to set the LogLevel or MetricsLevel parameters.

monitoringConfiguration_logLevel :: Lens' MonitoringConfiguration (Maybe LogLevel)

Describes the verbosity of the CloudWatch Logs for an application.

monitoringConfiguration_metricsLevel :: Lens' MonitoringConfiguration (Maybe MetricsLevel)

Describes the granularity of the CloudWatch Logs for an application. The Parallelism level is not recommended for applications with a Parallelism over 64 due to excessive costs.

monitoringConfiguration_configurationType :: Lens' MonitoringConfiguration ConfigurationType

Describes whether to use the default CloudWatch logging configuration for an application. You must set this property to CUSTOM in order to set the LogLevel or MetricsLevel parameters.