Amazonka.KinesisAnalyticsV2.Types.OutputDescription

Description

Documentation

data OutputDescription

For a SQL-based Kinesis Data Analytics application, describes the application output configuration, which includes the in-application stream name and the destination where the stream data is written. The destination can be a Kinesis data stream or a Kinesis Data Firehose delivery stream.

See: newOutputDescription smart constructor.

Constructors

OutputDescription'

Fields

outputId :: Maybe Text

A unique identifier for the output configuration.

destinationSchema :: Maybe DestinationSchema

The data format used for writing data to the destination.

kinesisFirehoseOutputDescription :: Maybe KinesisFirehoseOutputDescription

Describes the Kinesis Data Firehose delivery stream that is configured as the destination where output is written.

kinesisStreamsOutputDescription :: Maybe KinesisStreamsOutputDescription

Describes the Kinesis data stream that is configured as the destination where output is written.

name :: Maybe Text

The name of the in-application stream that is configured as output.

lambdaOutputDescription :: Maybe LambdaOutputDescription

Describes the Lambda function that is configured as the destination where output is written.

newOutputDescription :: OutputDescription

Create a value of OutputDescription with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:outputId:OutputDescription', outputDescription_outputId - A unique identifier for the output configuration.

$sel:destinationSchema:OutputDescription', outputDescription_destinationSchema - The data format used for writing data to the destination.

$sel:kinesisFirehoseOutputDescription:OutputDescription', outputDescription_kinesisFirehoseOutputDescription - Describes the Kinesis Data Firehose delivery stream that is configured as the destination where output is written.

$sel:kinesisStreamsOutputDescription:OutputDescription', outputDescription_kinesisStreamsOutputDescription - Describes the Kinesis data stream that is configured as the destination where output is written.

$sel:name:OutputDescription', outputDescription_name - The name of the in-application stream that is configured as output.

$sel:lambdaOutputDescription:OutputDescription', outputDescription_lambdaOutputDescription - Describes the Lambda function that is configured as the destination where output is written.

outputDescription_outputId :: Lens' OutputDescription (Maybe Text)

A unique identifier for the output configuration.

outputDescription_destinationSchema :: Lens' OutputDescription (Maybe DestinationSchema)

The data format used for writing data to the destination.

outputDescription_kinesisFirehoseOutputDescription :: Lens' OutputDescription (Maybe KinesisFirehoseOutputDescription)

Describes the Kinesis Data Firehose delivery stream that is configured as the destination where output is written.

outputDescription_kinesisStreamsOutputDescription :: Lens' OutputDescription (Maybe KinesisStreamsOutputDescription)

Describes the Kinesis data stream that is configured as the destination where output is written.

outputDescription_name :: Lens' OutputDescription (Maybe Text)

The name of the in-application stream that is configured as output.

outputDescription_lambdaOutputDescription :: Lens' OutputDescription (Maybe LambdaOutputDescription)

Describes the Lambda function that is configured as the destination where output is written.