Amazonka.KinesisAnalyticsV2.Types.S3ContentLocationUpdate

Description

Documentation

data S3ContentLocationUpdate

Describes an update for the Amazon S3 code content location for an application.

See: newS3ContentLocationUpdate smart constructor.

Constructors

S3ContentLocationUpdate'

Fields

bucketARNUpdate :: Maybe Text

The new Amazon Resource Name (ARN) for the S3 bucket containing the application code.

fileKeyUpdate :: Maybe Text

The new file key for the object containing the application code.

objectVersionUpdate :: Maybe Text

The new version of the object containing the application code.

newS3ContentLocationUpdate :: S3ContentLocationUpdate

Create a value of S3ContentLocationUpdate with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:bucketARNUpdate:S3ContentLocationUpdate', s3ContentLocationUpdate_bucketARNUpdate - The new Amazon Resource Name (ARN) for the S3 bucket containing the application code.

$sel:fileKeyUpdate:S3ContentLocationUpdate', s3ContentLocationUpdate_fileKeyUpdate - The new file key for the object containing the application code.

$sel:objectVersionUpdate:S3ContentLocationUpdate', s3ContentLocationUpdate_objectVersionUpdate - The new version of the object containing the application code.

s3ContentLocationUpdate_bucketARNUpdate :: Lens' S3ContentLocationUpdate (Maybe Text)

The new Amazon Resource Name (ARN) for the S3 bucket containing the application code.

s3ContentLocationUpdate_fileKeyUpdate :: Lens' S3ContentLocationUpdate (Maybe Text)

The new file key for the object containing the application code.

s3ContentLocationUpdate_objectVersionUpdate :: Lens' S3ContentLocationUpdate (Maybe Text)

The new version of the object containing the application code.