Amazonka.KinesisAnalyticsV2.Types.SqlApplicationConfiguration

Description

Documentation

data SqlApplicationConfiguration

Describes the inputs, outputs, and reference data sources for a SQL-based Kinesis Data Analytics application.

See: newSqlApplicationConfiguration smart constructor.

Constructors

SqlApplicationConfiguration'

Fields

inputs :: Maybe [Input]

The array of Input objects describing the input streams used by the application.

referenceDataSources :: Maybe [ReferenceDataSource]

The array of ReferenceDataSource objects describing the reference data sources used by the application.

outputs :: Maybe [Output]

The array of Output objects describing the destination streams used by the application.

newSqlApplicationConfiguration :: SqlApplicationConfiguration

Create a value of SqlApplicationConfiguration with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:inputs:SqlApplicationConfiguration', sqlApplicationConfiguration_inputs - The array of Input objects describing the input streams used by the application.

$sel:referenceDataSources:SqlApplicationConfiguration', sqlApplicationConfiguration_referenceDataSources - The array of ReferenceDataSource objects describing the reference data sources used by the application.

$sel:outputs:SqlApplicationConfiguration', sqlApplicationConfiguration_outputs - The array of Output objects describing the destination streams used by the application.

sqlApplicationConfiguration_inputs :: Lens' SqlApplicationConfiguration (Maybe [Input])

The array of Input objects describing the input streams used by the application.

sqlApplicationConfiguration_referenceDataSources :: Lens' SqlApplicationConfiguration (Maybe [ReferenceDataSource])

The array of ReferenceDataSource objects describing the reference data sources used by the application.

sqlApplicationConfiguration_outputs :: Lens' SqlApplicationConfiguration (Maybe [Output])

The array of Output objects describing the destination streams used by the application.