Amazonka.KinesisVideo

Description

Derived from API version 2017-09-30 of the AWS service descriptions, licensed under Apache 2.0.

Service Configuration

defaultService :: Service

API version 2017-09-30 of the Amazon Kinesis Video Streams SDK configuration.

Errors

Error matchers are designed for use with the functions provided by Control.Exception.Lens. This allows catching (and rethrowing) service specific errors returned by KinesisVideo.

AccessDeniedException

_AccessDeniedException :: AsError a => Getting (First ServiceError) a ServiceError

You do not have required permissions to perform this operation.

InvalidArgumentException

_InvalidArgumentException :: AsError a => Getting (First ServiceError) a ServiceError

The value for this input parameter is invalid.

TagsPerResourceExceededLimitException

_TagsPerResourceExceededLimitException :: AsError a => Getting (First ServiceError) a ServiceError

You have exceeded the limit of tags that you can associate with the resource. Kinesis video streams support up to 50 tags.

NotAuthorizedException

_NotAuthorizedException :: AsError a => Getting (First ServiceError) a ServiceError

The caller is not authorized to perform this operation.

ClientLimitExceededException

_ClientLimitExceededException :: AsError a => Getting (First ServiceError) a ServiceError

Kinesis Video Streams has throttled the request because you have exceeded the limit of allowed client calls. Try making the call later.

AccountChannelLimitExceededException

_AccountChannelLimitExceededException :: AsError a => Getting (First ServiceError) a ServiceError

You have reached the maximum limit of active signaling channels for this AWS account in this region.

InvalidDeviceException

_InvalidDeviceException :: AsError a => Getting (First ServiceError) a ServiceError

Not implemented.

VersionMismatchException

_VersionMismatchException :: AsError a => Getting (First ServiceError) a ServiceError

The stream version that you specified is not the latest version. To get the latest version, use the DescribeStream API.

AccountStreamLimitExceededException

_AccountStreamLimitExceededException :: AsError a => Getting (First ServiceError) a ServiceError

The number of streams created for the account is too high.

InvalidResourceFormatException

_InvalidResourceFormatException :: AsError a => Getting (First ServiceError) a ServiceError

The format of the StreamARN is invalid.

DeviceStreamLimitExceededException

_DeviceStreamLimitExceededException :: AsError a => Getting (First ServiceError) a ServiceError

Not implemented.

ResourceNotFoundException

_ResourceNotFoundException :: AsError a => Getting (First ServiceError) a ServiceError

Amazon Kinesis Video Streams can't find the stream that you specified.

ResourceInUseException

_ResourceInUseException :: AsError a => Getting (First ServiceError) a ServiceError

The signaling channel is currently not available for this operation.

Waiters

Waiters poll by repeatedly sending a request until some remote success condition configured by the Wait specification is fulfilled. The Wait specification determines how many attempts should be made, in addition to delay and retry strategies.

Operations

Some AWS operations return results that are incomplete and require subsequent requests in order to obtain the entire result set. The process of sending subsequent requests to continue where a previous request left off is called pagination. For example, the ListObjects operation of Amazon S3 returns up to 1000 objects at a time, and you must send subsequent requests with the appropriate Marker in order to retrieve the next page of results.

Operations that have an AWSPager instance can transparently perform subsequent requests, correctly setting Markers and other request facets to iterate through the entire result set of a truncated API operation. Operations which support this have an additional note in the documentation.

Many operations have the ability to filter results on the server side. See the individual operation parameters for details.

ListSignalingChannels (Paginated)

data ListSignalingChannels

See: newListSignalingChannels smart constructor.

Constructors

ListSignalingChannels' (Maybe ChannelNameCondition) (Maybe Text) (Maybe Natural)

newListSignalingChannels :: ListSignalingChannels

Create a value of ListSignalingChannels with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:channelNameCondition:ListSignalingChannels', listSignalingChannels_channelNameCondition - Optional: Returns only the channels that satisfy a specific condition.

$sel:nextToken:ListSignalingChannels', listSignalingChannels_nextToken - If you specify this parameter, when the result of a ListSignalingChannels operation is truncated, the call returns the NextToken in the response. To get another batch of channels, provide this token in your next request.

$sel:maxResults:ListSignalingChannels', listSignalingChannels_maxResults - The maximum number of channels to return in the response. The default is 500.

data ListSignalingChannelsResponse

See: newListSignalingChannelsResponse smart constructor.

Constructors

ListSignalingChannelsResponse' (Maybe [ChannelInfo]) (Maybe Text) Int

newListSignalingChannelsResponse

Arguments

:: Int
$sel:httpStatus:ListSignalingChannelsResponse'

-> ListSignalingChannelsResponse

Create a value of ListSignalingChannelsResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:channelInfoList:ListSignalingChannelsResponse', listSignalingChannelsResponse_channelInfoList - An array of ChannelInfo objects.

$sel:nextToken:ListSignalingChannels', listSignalingChannelsResponse_nextToken - If the response is truncated, the call returns this element with a token. To get the next batch of streams, use this token in your next request.

$sel:httpStatus:ListSignalingChannelsResponse', listSignalingChannelsResponse_httpStatus - The response's http status code.

ListTagsForResource

data ListTagsForResource

See: newListTagsForResource smart constructor.

Constructors

ListTagsForResource' (Maybe Text) Text

newListTagsForResource

Arguments

:: Text
$sel:resourceARN:ListTagsForResource'

-> ListTagsForResource

Create a value of ListTagsForResource with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:nextToken:ListTagsForResource', listTagsForResource_nextToken - If you specify this parameter and the result of a ListTagsForResource call is truncated, the response includes a token that you can use in the next request to fetch the next batch of tags.

$sel:resourceARN:ListTagsForResource', listTagsForResource_resourceARN - The Amazon Resource Name (ARN) of the signaling channel for which you want to list tags.

data ListTagsForResourceResponse

See: newListTagsForResourceResponse smart constructor.

Constructors

ListTagsForResourceResponse' (Maybe Text) (Maybe (HashMap Text Text)) Int

newListTagsForResourceResponse

Arguments

:: Int
$sel:httpStatus:ListTagsForResourceResponse'

-> ListTagsForResourceResponse

Create a value of ListTagsForResourceResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:nextToken:ListTagsForResource', listTagsForResourceResponse_nextToken - If you specify this parameter and the result of a ListTagsForResource call is truncated, the response includes a token that you can use in the next request to fetch the next set of tags.

$sel:tags:ListTagsForResourceResponse', listTagsForResourceResponse_tags - A map of tag keys and values associated with the specified signaling channel.

$sel:httpStatus:ListTagsForResourceResponse', listTagsForResourceResponse_httpStatus - The response's http status code.

UntagStream

data UntagStream

See: newUntagStream smart constructor.

Constructors

UntagStream' (Maybe Text) (Maybe Text) (NonEmpty Text)

newUntagStream

Arguments

:: NonEmpty Text
$sel:tagKeyList:UntagStream'

-> UntagStream

Create a value of UntagStream with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:streamARN:UntagStream', untagStream_streamARN - The Amazon Resource Name (ARN) of the stream that you want to remove tags from.

$sel:streamName:UntagStream', untagStream_streamName - The name of the stream that you want to remove tags from.

$sel:tagKeyList:UntagStream', untagStream_tagKeyList - A list of the keys of the tags that you want to remove.

data UntagStreamResponse

See: newUntagStreamResponse smart constructor.

Constructors

UntagStreamResponse' Int

newUntagStreamResponse

Arguments

:: Int
$sel:httpStatus:UntagStreamResponse'

-> UntagStreamResponse

Create a value of UntagStreamResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:httpStatus:UntagStreamResponse', untagStreamResponse_httpStatus - The response's http status code.

UpdateDataRetention

data UpdateDataRetention

See: newUpdateDataRetention smart constructor.

Constructors

UpdateDataRetention' (Maybe Text) (Maybe Text) Text UpdateDataRetentionOperation Natural

newUpdateDataRetention

Arguments

:: Text
$sel:currentVersion:UpdateDataRetention'

-> UpdateDataRetentionOperation
$sel:operation:UpdateDataRetention'

-> Natural
$sel:dataRetentionChangeInHours:UpdateDataRetention'

-> UpdateDataRetention

Create a value of UpdateDataRetention with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:streamARN:UpdateDataRetention', updateDataRetention_streamARN - The Amazon Resource Name (ARN) of the stream whose retention period you want to change.

$sel:streamName:UpdateDataRetention', updateDataRetention_streamName - The name of the stream whose retention period you want to change.

$sel:currentVersion:UpdateDataRetention', updateDataRetention_currentVersion - The version of the stream whose retention period you want to change. To get the version, call either the DescribeStream or the ListStreams API.

$sel:operation:UpdateDataRetention', updateDataRetention_operation - Indicates whether you want to increase or decrease the retention period.

$sel:dataRetentionChangeInHours:UpdateDataRetention', updateDataRetention_dataRetentionChangeInHours - The retention period, in hours. The value you specify replaces the current value. The maximum value for this parameter is 87600 (ten years).

data UpdateDataRetentionResponse

See: newUpdateDataRetentionResponse smart constructor.

Constructors

UpdateDataRetentionResponse' Int

newUpdateDataRetentionResponse

Arguments

:: Int
$sel:httpStatus:UpdateDataRetentionResponse'

-> UpdateDataRetentionResponse

Create a value of UpdateDataRetentionResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:httpStatus:UpdateDataRetentionResponse', updateDataRetentionResponse_httpStatus - The response's http status code.

GetDataEndpoint

data GetDataEndpoint

See: newGetDataEndpoint smart constructor.

Constructors

GetDataEndpoint' (Maybe Text) (Maybe Text) APIName

newGetDataEndpoint

Arguments

:: APIName
$sel:aPIName:GetDataEndpoint'

-> GetDataEndpoint

Create a value of GetDataEndpoint with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:streamARN:GetDataEndpoint', getDataEndpoint_streamARN - The Amazon Resource Name (ARN) of the stream that you want to get the endpoint for. You must specify either this parameter or a StreamName in the request.

$sel:streamName:GetDataEndpoint', getDataEndpoint_streamName - The name of the stream that you want to get the endpoint for. You must specify either this parameter or a StreamARN in the request.

$sel:aPIName:GetDataEndpoint', getDataEndpoint_aPIName - The name of the API action for which to get an endpoint.

data GetDataEndpointResponse

See: newGetDataEndpointResponse smart constructor.

Constructors

GetDataEndpointResponse' (Maybe Text) Int

newGetDataEndpointResponse

Arguments

:: Int
$sel:httpStatus:GetDataEndpointResponse'

-> GetDataEndpointResponse

Create a value of GetDataEndpointResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:dataEndpoint:GetDataEndpointResponse', getDataEndpointResponse_dataEndpoint - The endpoint value. To read data from the stream or to write data to it, specify this endpoint in your application.

$sel:httpStatus:GetDataEndpointResponse', getDataEndpointResponse_httpStatus - The response's http status code.

GetSignalingChannelEndpoint

data GetSignalingChannelEndpoint

See: newGetSignalingChannelEndpoint smart constructor.

Constructors

GetSignalingChannelEndpoint' (Maybe SingleMasterChannelEndpointConfiguration) Text

newGetSignalingChannelEndpoint

Arguments

:: Text
$sel:channelARN:GetSignalingChannelEndpoint'

-> GetSignalingChannelEndpoint

Create a value of GetSignalingChannelEndpoint with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:singleMasterChannelEndpointConfiguration:GetSignalingChannelEndpoint', getSignalingChannelEndpoint_singleMasterChannelEndpointConfiguration - A structure containing the endpoint configuration for the SINGLE_MASTER channel type.

$sel:channelARN:GetSignalingChannelEndpoint', getSignalingChannelEndpoint_channelARN - The Amazon Resource Name (ARN) of the signalling channel for which you want to get an endpoint.

data GetSignalingChannelEndpointResponse

See: newGetSignalingChannelEndpointResponse smart constructor.

Constructors

GetSignalingChannelEndpointResponse' (Maybe [ResourceEndpointListItem]) Int

newGetSignalingChannelEndpointResponse

Arguments

:: Int
$sel:httpStatus:GetSignalingChannelEndpointResponse'

-> GetSignalingChannelEndpointResponse

Create a value of GetSignalingChannelEndpointResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:resourceEndpointList:GetSignalingChannelEndpointResponse', getSignalingChannelEndpointResponse_resourceEndpointList - A list of endpoints for the specified signaling channel.

$sel:httpStatus:GetSignalingChannelEndpointResponse', getSignalingChannelEndpointResponse_httpStatus - The response's http status code.

ListTagsForStream

data ListTagsForStream

See: newListTagsForStream smart constructor.

Constructors

ListTagsForStream' (Maybe Text) (Maybe Text) (Maybe Text)

newListTagsForStream :: ListTagsForStream

Create a value of ListTagsForStream with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:streamARN:ListTagsForStream', listTagsForStream_streamARN - The Amazon Resource Name (ARN) of the stream that you want to list tags for.

$sel:nextToken:ListTagsForStream', listTagsForStream_nextToken - If you specify this parameter and the result of a ListTagsForStream call is truncated, the response includes a token that you can use in the next request to fetch the next batch of tags.

$sel:streamName:ListTagsForStream', listTagsForStream_streamName - The name of the stream that you want to list tags for.

data ListTagsForStreamResponse

See: newListTagsForStreamResponse smart constructor.

Constructors

ListTagsForStreamResponse' (Maybe Text) (Maybe (HashMap Text Text)) Int

newListTagsForStreamResponse

Arguments

:: Int
$sel:httpStatus:ListTagsForStreamResponse'

-> ListTagsForStreamResponse

Create a value of ListTagsForStreamResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:nextToken:ListTagsForStream', listTagsForStreamResponse_nextToken - If you specify this parameter and the result of a ListTags call is truncated, the response includes a token that you can use in the next request to fetch the next set of tags.

$sel:tags:ListTagsForStreamResponse', listTagsForStreamResponse_tags - A map of tag keys and values associated with the specified stream.

$sel:httpStatus:ListTagsForStreamResponse', listTagsForStreamResponse_httpStatus - The response's http status code.

DeleteSignalingChannel

data DeleteSignalingChannel

See: newDeleteSignalingChannel smart constructor.

Constructors

DeleteSignalingChannel' (Maybe Text) Text

newDeleteSignalingChannel

Arguments

:: Text
$sel:channelARN:DeleteSignalingChannel'

-> DeleteSignalingChannel

Create a value of DeleteSignalingChannel with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:currentVersion:DeleteSignalingChannel', deleteSignalingChannel_currentVersion - The current version of the signaling channel that you want to delete. You can obtain the current version by invoking the DescribeSignalingChannel or ListSignalingChannels API operations.

$sel:channelARN:DeleteSignalingChannel', deleteSignalingChannel_channelARN - The Amazon Resource Name (ARN) of the signaling channel that you want to delete.

data DeleteSignalingChannelResponse

See: newDeleteSignalingChannelResponse smart constructor.

Constructors

DeleteSignalingChannelResponse' Int

newDeleteSignalingChannelResponse

Arguments

:: Int
$sel:httpStatus:DeleteSignalingChannelResponse'

-> DeleteSignalingChannelResponse

Create a value of DeleteSignalingChannelResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:httpStatus:DeleteSignalingChannelResponse', deleteSignalingChannelResponse_httpStatus - The response's http status code.

UpdateSignalingChannel

data UpdateSignalingChannel

See: newUpdateSignalingChannel smart constructor.

Constructors

UpdateSignalingChannel' (Maybe SingleMasterConfiguration) Text Text

newUpdateSignalingChannel

Arguments

:: Text
$sel:channelARN:UpdateSignalingChannel'

-> Text
$sel:currentVersion:UpdateSignalingChannel'

-> UpdateSignalingChannel

Create a value of UpdateSignalingChannel with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:singleMasterConfiguration:UpdateSignalingChannel', updateSignalingChannel_singleMasterConfiguration - The structure containing the configuration for the SINGLE_MASTER type of the signaling channel that you want to update.

$sel:channelARN:UpdateSignalingChannel', updateSignalingChannel_channelARN - The Amazon Resource Name (ARN) of the signaling channel that you want to update.

$sel:currentVersion:UpdateSignalingChannel', updateSignalingChannel_currentVersion - The current version of the signaling channel that you want to update.

data UpdateSignalingChannelResponse

See: newUpdateSignalingChannelResponse smart constructor.

Constructors

UpdateSignalingChannelResponse' Int

newUpdateSignalingChannelResponse

Arguments

:: Int
$sel:httpStatus:UpdateSignalingChannelResponse'

-> UpdateSignalingChannelResponse

Create a value of UpdateSignalingChannelResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:httpStatus:UpdateSignalingChannelResponse', updateSignalingChannelResponse_httpStatus - The response's http status code.

UpdateStream

data UpdateStream

See: newUpdateStream smart constructor.

Constructors

UpdateStream' (Maybe Text) (Maybe Text) (Maybe Text) (Maybe Text) Text

newUpdateStream

Arguments

:: Text
$sel:currentVersion:UpdateStream'

-> UpdateStream

Create a value of UpdateStream with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:mediaType:UpdateStream', updateStream_mediaType - The stream's media type. Use MediaType to specify the type of content that the stream contains to the consumers of the stream. For more information about media types, see Media Types. If you choose to specify the MediaType, see Naming Requirements.

To play video on the console, you must specify the correct video type. For example, if the video in the stream is H.264, specify video/h264 as the MediaType.

$sel:streamARN:UpdateStream', updateStream_streamARN - The ARN of the stream whose metadata you want to update.

$sel:deviceName:UpdateStream', updateStream_deviceName - The name of the device that is writing to the stream.

In the current implementation, Kinesis Video Streams does not use this name.

$sel:streamName:UpdateStream', updateStream_streamName - The name of the stream whose metadata you want to update.

The stream name is an identifier for the stream, and must be unique for each account and region.

$sel:currentVersion:UpdateStream', updateStream_currentVersion - The version of the stream whose metadata you want to update.

data UpdateStreamResponse

See: newUpdateStreamResponse smart constructor.

Constructors

UpdateStreamResponse' Int

newUpdateStreamResponse

Arguments

:: Int
$sel:httpStatus:UpdateStreamResponse'

-> UpdateStreamResponse

Create a value of UpdateStreamResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:httpStatus:UpdateStreamResponse', updateStreamResponse_httpStatus - The response's http status code.

DeleteStream

data DeleteStream

See: newDeleteStream smart constructor.

Constructors

DeleteStream' (Maybe Text) Text

newDeleteStream

Arguments

:: Text
$sel:streamARN:DeleteStream'

-> DeleteStream

Create a value of DeleteStream with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:currentVersion:DeleteStream', deleteStream_currentVersion - Optional: The version of the stream that you want to delete.

Specify the version as a safeguard to ensure that your are deleting the correct stream. To get the stream version, use the DescribeStream API.

If not specified, only the CreationTime is checked before deleting the stream.

$sel:streamARN:DeleteStream', deleteStream_streamARN - The Amazon Resource Name (ARN) of the stream that you want to delete.

data DeleteStreamResponse

See: newDeleteStreamResponse smart constructor.

Constructors

DeleteStreamResponse' Int

newDeleteStreamResponse

Arguments

:: Int
$sel:httpStatus:DeleteStreamResponse'

-> DeleteStreamResponse

Create a value of DeleteStreamResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:httpStatus:DeleteStreamResponse', deleteStreamResponse_httpStatus - The response's http status code.

ListStreams (Paginated)

data ListStreams

See: newListStreams smart constructor.

Constructors

ListStreams' (Maybe Text) (Maybe StreamNameCondition) (Maybe Natural)

newListStreams :: ListStreams

Create a value of ListStreams with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:nextToken:ListStreams', listStreams_nextToken - If you specify this parameter, when the result of a ListStreams operation is truncated, the call returns the NextToken in the response. To get another batch of streams, provide this token in your next request.

$sel:streamNameCondition:ListStreams', listStreams_streamNameCondition - Optional: Returns only streams that satisfy a specific condition. Currently, you can specify only the prefix of a stream name as a condition.

$sel:maxResults:ListStreams', listStreams_maxResults - The maximum number of streams to return in the response. The default is 10,000.

data ListStreamsResponse

See: newListStreamsResponse smart constructor.

Constructors

ListStreamsResponse' (Maybe [StreamInfo]) (Maybe Text) Int

newListStreamsResponse

Arguments

:: Int
$sel:httpStatus:ListStreamsResponse'

-> ListStreamsResponse

Create a value of ListStreamsResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:streamInfoList:ListStreamsResponse', listStreamsResponse_streamInfoList - An array of StreamInfo objects.

$sel:nextToken:ListStreams', listStreamsResponse_nextToken - If the response is truncated, the call returns this element with a token. To get the next batch of streams, use this token in your next request.

$sel:httpStatus:ListStreamsResponse', listStreamsResponse_httpStatus - The response's http status code.

CreateStream

data CreateStream

See: newCreateStream smart constructor.

Constructors

CreateStream' (Maybe Text) (Maybe Natural) (Maybe Text) (Maybe Text) (Maybe (HashMap Text Text)) Text

newCreateStream

Arguments

:: Text
$sel:streamName:CreateStream'

-> CreateStream

Create a value of CreateStream with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:mediaType:CreateStream', createStream_mediaType - The media type of the stream. Consumers of the stream can use this information when processing the stream. For more information about media types, see Media Types. If you choose to specify the MediaType, see Naming Requirements for guidelines.

Example valid values include "video/h264" and "video/h264,audio/aac".

This parameter is optional; the default value is null (or empty in JSON).

$sel:dataRetentionInHours:CreateStream', createStream_dataRetentionInHours - The number of hours that you want to retain the data in the stream. Kinesis Video Streams retains the data in a data store that is associated with the stream.

The default value is 0, indicating that the stream does not persist data.

When the DataRetentionInHours value is 0, consumers can still consume the fragments that remain in the service host buffer, which has a retention time limit of 5 minutes and a retention memory limit of 200 MB. Fragments are removed from the buffer when either limit is reached.

$sel:kmsKeyId:CreateStream', createStream_kmsKeyId - The ID of the AWS Key Management Service (AWS KMS) key that you want Kinesis Video Streams to use to encrypt stream data.

If no key ID is specified, the default, Kinesis Video-managed key (aws/kinesisvideo) is used.

For more information, see DescribeKey.

$sel:deviceName:CreateStream', createStream_deviceName - The name of the device that is writing to the stream.

In the current implementation, Kinesis Video Streams does not use this name.

$sel:tags:CreateStream', createStream_tags - A list of tags to associate with the specified stream. Each tag is a key-value pair (the value is optional).

$sel:streamName:CreateStream', createStream_streamName - A name for the stream that you are creating.

The stream name is an identifier for the stream, and must be unique for each account and region.

data CreateStreamResponse

See: newCreateStreamResponse smart constructor.

Constructors

CreateStreamResponse' (Maybe Text) Int

newCreateStreamResponse

Arguments

:: Int
$sel:httpStatus:CreateStreamResponse'

-> CreateStreamResponse

Create a value of CreateStreamResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:streamARN:CreateStreamResponse', createStreamResponse_streamARN - The Amazon Resource Name (ARN) of the stream.

$sel:httpStatus:CreateStreamResponse', createStreamResponse_httpStatus - The response's http status code.

DescribeSignalingChannel

data DescribeSignalingChannel

See: newDescribeSignalingChannel smart constructor.

Constructors

DescribeSignalingChannel' (Maybe Text) (Maybe Text)

newDescribeSignalingChannel :: DescribeSignalingChannel

Create a value of DescribeSignalingChannel with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:channelARN:DescribeSignalingChannel', describeSignalingChannel_channelARN - The ARN of the signaling channel that you want to describe.

$sel:channelName:DescribeSignalingChannel', describeSignalingChannel_channelName - The name of the signaling channel that you want to describe.

data DescribeSignalingChannelResponse

See: newDescribeSignalingChannelResponse smart constructor.

Constructors

DescribeSignalingChannelResponse' (Maybe ChannelInfo) Int

newDescribeSignalingChannelResponse

Arguments

:: Int
$sel:httpStatus:DescribeSignalingChannelResponse'

-> DescribeSignalingChannelResponse

Create a value of DescribeSignalingChannelResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:channelInfo:DescribeSignalingChannelResponse', describeSignalingChannelResponse_channelInfo - A structure that encapsulates the specified signaling channel's metadata and properties.

$sel:httpStatus:DescribeSignalingChannelResponse', describeSignalingChannelResponse_httpStatus - The response's http status code.

TagResource

data TagResource

See: newTagResource smart constructor.

Constructors

TagResource' Text (NonEmpty Tag)

newTagResource

Arguments

:: Text
$sel:resourceARN:TagResource'

-> NonEmpty Tag
$sel:tags:TagResource'

-> TagResource

Create a value of TagResource with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:resourceARN:TagResource', tagResource_resourceARN - The Amazon Resource Name (ARN) of the signaling channel to which you want to add tags.

$sel:tags:TagResource', tagResource_tags - A list of tags to associate with the specified signaling channel. Each tag is a key-value pair.

data TagResourceResponse

See: newTagResourceResponse smart constructor.

Constructors

TagResourceResponse' Int

newTagResourceResponse

Arguments

:: Int
$sel:httpStatus:TagResourceResponse'

-> TagResourceResponse

Create a value of TagResourceResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:httpStatus:TagResourceResponse', tagResourceResponse_httpStatus - The response's http status code.

UntagResource

data UntagResource

See: newUntagResource smart constructor.

Constructors

UntagResource' Text (NonEmpty Text)

newUntagResource

Arguments

:: Text
$sel:resourceARN:UntagResource'

-> NonEmpty Text
$sel:tagKeyList:UntagResource'

-> UntagResource

Create a value of UntagResource with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:resourceARN:UntagResource', untagResource_resourceARN - The Amazon Resource Name (ARN) of the signaling channel from which you want to remove tags.

$sel:tagKeyList:UntagResource', untagResource_tagKeyList - A list of the keys of the tags that you want to remove.

data UntagResourceResponse

See: newUntagResourceResponse smart constructor.

Constructors

UntagResourceResponse' Int

newUntagResourceResponse

Arguments

:: Int
$sel:httpStatus:UntagResourceResponse'

-> UntagResourceResponse

Create a value of UntagResourceResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:httpStatus:UntagResourceResponse', untagResourceResponse_httpStatus - The response's http status code.

TagStream

data TagStream

See: newTagStream smart constructor.

Constructors

TagStream' (Maybe Text) (Maybe Text) (HashMap Text Text)

newTagStream :: TagStream

Create a value of TagStream with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:streamARN:TagStream', tagStream_streamARN - The Amazon Resource Name (ARN) of the resource that you want to add the tag or tags to.

$sel:streamName:TagStream', tagStream_streamName - The name of the stream that you want to add the tag or tags to.

$sel:tags:TagStream', tagStream_tags - A list of tags to associate with the specified stream. Each tag is a key-value pair (the value is optional).

data TagStreamResponse

See: newTagStreamResponse smart constructor.

Constructors

TagStreamResponse' Int

newTagStreamResponse

Arguments

:: Int
$sel:httpStatus:TagStreamResponse'

-> TagStreamResponse

Create a value of TagStreamResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:httpStatus:TagStreamResponse', tagStreamResponse_httpStatus - The response's http status code.

CreateSignalingChannel

data CreateSignalingChannel

See: newCreateSignalingChannel smart constructor.

Constructors

CreateSignalingChannel' (Maybe SingleMasterConfiguration) (Maybe ChannelType) (Maybe [Tag]) Text

newCreateSignalingChannel

Arguments

:: Text
$sel:channelName:CreateSignalingChannel'

-> CreateSignalingChannel

Create a value of CreateSignalingChannel with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:singleMasterConfiguration:CreateSignalingChannel', createSignalingChannel_singleMasterConfiguration - A structure containing the configuration for the SINGLE_MASTER channel type.

$sel:channelType:CreateSignalingChannel', createSignalingChannel_channelType - A type of the signaling channel that you are creating. Currently, SINGLE_MASTER is the only supported channel type.

$sel:tags:CreateSignalingChannel', createSignalingChannel_tags - A set of tags (key-value pairs) that you want to associate with this channel.

$sel:channelName:CreateSignalingChannel', createSignalingChannel_channelName - A name for the signaling channel that you are creating. It must be unique for each AWS account and AWS Region.

data CreateSignalingChannelResponse

See: newCreateSignalingChannelResponse smart constructor.

Constructors

CreateSignalingChannelResponse' (Maybe Text) Int

newCreateSignalingChannelResponse

Arguments

:: Int
$sel:httpStatus:CreateSignalingChannelResponse'

-> CreateSignalingChannelResponse

Create a value of CreateSignalingChannelResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:channelARN:CreateSignalingChannelResponse', createSignalingChannelResponse_channelARN - The Amazon Resource Name (ARN) of the created channel.

$sel:httpStatus:CreateSignalingChannelResponse', createSignalingChannelResponse_httpStatus - The response's http status code.

DescribeStream

data DescribeStream

See: newDescribeStream smart constructor.

Constructors

DescribeStream' (Maybe Text) (Maybe Text)

newDescribeStream :: DescribeStream

Create a value of DescribeStream with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:streamARN:DescribeStream', describeStream_streamARN - The Amazon Resource Name (ARN) of the stream.

$sel:streamName:DescribeStream', describeStream_streamName - The name of the stream.

data DescribeStreamResponse

See: newDescribeStreamResponse smart constructor.

Constructors

DescribeStreamResponse' (Maybe StreamInfo) Int

newDescribeStreamResponse

Arguments

:: Int
$sel:httpStatus:DescribeStreamResponse'

-> DescribeStreamResponse

Create a value of DescribeStreamResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:streamInfo:DescribeStreamResponse', describeStreamResponse_streamInfo - An object that describes the stream.

$sel:httpStatus:DescribeStreamResponse', describeStreamResponse_httpStatus - The response's http status code.

Types

APIName

newtype APIName

Constructors

APIName'

Fields

fromAPIName :: Text

Bundled Patterns

pattern APIName_GET_CLIP :: APIName

pattern APIName_GET_DASH_STREAMING_SESSION_URL :: APIName

pattern APIName_GET_HLS_STREAMING_SESSION_URL :: APIName

pattern APIName_GET_MEDIA :: APIName

pattern APIName_GET_MEDIA_FOR_FRAGMENT_LIST :: APIName

pattern APIName_LIST_FRAGMENTS :: APIName

pattern APIName_PUT_MEDIA :: APIName

ChannelProtocol

newtype ChannelProtocol

Constructors

ChannelProtocol'

Fields

fromChannelProtocol :: Text

Bundled Patterns

pattern ChannelProtocol_HTTPS :: ChannelProtocol

pattern ChannelProtocol_WSS :: ChannelProtocol

ChannelRole

newtype ChannelRole

Constructors

ChannelRole'

Fields

fromChannelRole :: Text

Bundled Patterns

pattern ChannelRole_MASTER :: ChannelRole

pattern ChannelRole_VIEWER :: ChannelRole

ChannelType

newtype ChannelType

Constructors

ChannelType'

Fields

fromChannelType :: Text

Bundled Patterns

pattern ChannelType_SINGLE_MASTER :: ChannelType

ComparisonOperator

newtype ComparisonOperator

Constructors

ComparisonOperator'

Fields

fromComparisonOperator :: Text

Bundled Patterns

pattern ComparisonOperator_BEGINS_WITH :: ComparisonOperator

StreamStatus

newtype StreamStatus

Constructors

StreamStatus'

Fields

fromStreamStatus :: Text

Bundled Patterns

pattern StreamStatus_ACTIVE :: StreamStatus

pattern StreamStatus_CREATING :: StreamStatus

pattern StreamStatus_DELETING :: StreamStatus

pattern StreamStatus_UPDATING :: StreamStatus

UpdateDataRetentionOperation

newtype UpdateDataRetentionOperation

Constructors

UpdateDataRetentionOperation'

Fields

fromUpdateDataRetentionOperation :: Text

Bundled Patterns

pattern UpdateDataRetentionOperation_DECREASE_DATA_RETENTION :: UpdateDataRetentionOperation

pattern UpdateDataRetentionOperation_INCREASE_DATA_RETENTION :: UpdateDataRetentionOperation

ChannelInfo

data ChannelInfo

A structure that encapsulates a signaling channel's metadata and properties.

See: newChannelInfo smart constructor.

Constructors

ChannelInfo' (Maybe POSIX) (Maybe StreamStatus) (Maybe Text) (Maybe SingleMasterConfiguration) (Maybe Text) (Maybe Text) (Maybe ChannelType)

newChannelInfo :: ChannelInfo

Create a value of ChannelInfo with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:creationTime:ChannelInfo', channelInfo_creationTime - The time at which the signaling channel was created.

$sel:channelStatus:ChannelInfo', channelInfo_channelStatus - Current status of the signaling channel.

$sel:channelARN:ChannelInfo', channelInfo_channelARN - The Amazon Resource Name (ARN) of the signaling channel.

$sel:singleMasterConfiguration:ChannelInfo', channelInfo_singleMasterConfiguration - A structure that contains the configuration for the SINGLE_MASTER channel type.

$sel:channelName:ChannelInfo', channelInfo_channelName - The name of the signaling channel.

$sel:version:ChannelInfo', channelInfo_version - The current version of the signaling channel.

$sel:channelType:ChannelInfo', channelInfo_channelType - The type of the signaling channel.

ChannelNameCondition

data ChannelNameCondition

An optional input parameter for the ListSignalingChannels API. When this parameter is specified while invoking ListSignalingChannels, the API returns only the channels that satisfy a condition specified in ChannelNameCondition.

See: newChannelNameCondition smart constructor.

Constructors

ChannelNameCondition' (Maybe ComparisonOperator) (Maybe Text)

newChannelNameCondition :: ChannelNameCondition

Create a value of ChannelNameCondition with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:comparisonOperator:ChannelNameCondition', channelNameCondition_comparisonOperator - A comparison operator. Currently, you can only specify the BEGINS_WITH operator, which finds signaling channels whose names begin with a given prefix.

$sel:comparisonValue:ChannelNameCondition', channelNameCondition_comparisonValue - A value to compare.

ResourceEndpointListItem

data ResourceEndpointListItem

An object that describes the endpoint of the signaling channel returned by the GetSignalingChannelEndpoint API.

See: newResourceEndpointListItem smart constructor.

Constructors

ResourceEndpointListItem' (Maybe ChannelProtocol) (Maybe Text)

newResourceEndpointListItem :: ResourceEndpointListItem

Create a value of ResourceEndpointListItem with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:protocol:ResourceEndpointListItem', resourceEndpointListItem_protocol - The protocol of the signaling channel returned by the GetSignalingChannelEndpoint API.

$sel:resourceEndpoint:ResourceEndpointListItem', resourceEndpointListItem_resourceEndpoint - The endpoint of the signaling channel returned by the GetSignalingChannelEndpoint API.

SingleMasterChannelEndpointConfiguration

data SingleMasterChannelEndpointConfiguration

An object that contains the endpoint configuration for the SINGLE_MASTER channel type.

See: newSingleMasterChannelEndpointConfiguration smart constructor.

Constructors

SingleMasterChannelEndpointConfiguration' (Maybe (NonEmpty ChannelProtocol)) (Maybe ChannelRole)

newSingleMasterChannelEndpointConfiguration :: SingleMasterChannelEndpointConfiguration

Create a value of SingleMasterChannelEndpointConfiguration with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:protocols:SingleMasterChannelEndpointConfiguration', singleMasterChannelEndpointConfiguration_protocols - This property is used to determine the nature of communication over this SINGLE_MASTER signaling channel. If WSS is specified, this API returns a websocket endpoint. If HTTPS is specified, this API returns an HTTPS endpoint.

$sel:role':SingleMasterChannelEndpointConfiguration', singleMasterChannelEndpointConfiguration_role - This property is used to determine messaging permissions in this SINGLE_MASTER signaling channel. If MASTER is specified, this API returns an endpoint that a client can use to receive offers from and send answers to any of the viewers on this signaling channel. If VIEWER is specified, this API returns an endpoint that a client can use only to send offers to another MASTER client on this signaling channel.

SingleMasterConfiguration

data SingleMasterConfiguration

A structure that contains the configuration for the SINGLE_MASTER channel type.

See: newSingleMasterConfiguration smart constructor.

Constructors

SingleMasterConfiguration' (Maybe Natural)

newSingleMasterConfiguration :: SingleMasterConfiguration

Create a value of SingleMasterConfiguration with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:messageTtlSeconds:SingleMasterConfiguration', singleMasterConfiguration_messageTtlSeconds - The period of time a signaling channel retains underlivered messages before they are discarded.

StreamInfo

data StreamInfo

An object describing a Kinesis video stream.

See: newStreamInfo smart constructor.

Constructors

StreamInfo' (Maybe POSIX) (Maybe StreamStatus) (Maybe Text) (Maybe Natural) (Maybe Text) (Maybe Text) (Maybe Text) (Maybe Text) (Maybe Text)

newStreamInfo :: StreamInfo

Create a value of StreamInfo with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:creationTime:StreamInfo', streamInfo_creationTime - A time stamp that indicates when the stream was created.

$sel:status:StreamInfo', streamInfo_status - The status of the stream.

$sel:mediaType:StreamInfo', streamInfo_mediaType - The MediaType of the stream.

$sel:dataRetentionInHours:StreamInfo', streamInfo_dataRetentionInHours - How long the stream retains data, in hours.

$sel:streamARN:StreamInfo', streamInfo_streamARN - The Amazon Resource Name (ARN) of the stream.

$sel:kmsKeyId:StreamInfo', streamInfo_kmsKeyId - The ID of the AWS Key Management Service (AWS KMS) key that Kinesis Video Streams uses to encrypt data on the stream.

$sel:deviceName:StreamInfo', streamInfo_deviceName - The name of the device that is associated with the stream.

$sel:version:StreamInfo', streamInfo_version - The version of the stream.

$sel:streamName:StreamInfo', streamInfo_streamName - The name of the stream.

StreamNameCondition

data StreamNameCondition

Specifies the condition that streams must satisfy to be returned when you list streams (see the ListStreams API). A condition has a comparison operation and a value. Currently, you can specify only the BEGINS_WITH operator, which finds streams whose names start with a given prefix.

See: newStreamNameCondition smart constructor.

Constructors

StreamNameCondition' (Maybe ComparisonOperator) (Maybe Text)

newStreamNameCondition :: StreamNameCondition

Create a value of StreamNameCondition with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:comparisonOperator:StreamNameCondition', streamNameCondition_comparisonOperator - A comparison operator. Currently, you can specify only the BEGINS_WITH operator, which finds streams whose names start with a given prefix.

$sel:comparisonValue:StreamNameCondition', streamNameCondition_comparisonValue - A value to compare.

Tag

data Tag

A key and value pair that is associated with the specified signaling channel.

See: newTag smart constructor.

Constructors

Tag' Text Text

newTag

Arguments

:: Text
$sel:key:Tag'

-> Text
$sel:value:Tag'

-> Tag

Create a value of Tag with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:key:Tag', tag_key - The key of the tag that is associated with the specified signaling channel.

$sel:value:Tag', tag_value - The value of the tag that is associated with the specified signaling channel.