Amazonka.KinesisVideo.GetDataEndpoint

Description

Gets an endpoint for a specified stream for either reading or writing. Use this endpoint in your application to read from the specified stream (using the GetMedia or GetMediaForFragmentList operations) or write to it (using the PutMedia operation).

The returned endpoint does not have the API name appended. The client needs to add the API name to the returned endpoint.

In the request, specify the stream either by StreamName or StreamARN.

Creating a Request

data GetDataEndpoint

See: newGetDataEndpoint smart constructor.

Constructors

GetDataEndpoint'

Fields

streamARN :: Maybe Text

The Amazon Resource Name (ARN) of the stream that you want to get the endpoint for. You must specify either this parameter or a StreamName in the request.

streamName :: Maybe Text

The name of the stream that you want to get the endpoint for. You must specify either this parameter or a StreamARN in the request.

aPIName :: APIName

The name of the API action for which to get an endpoint.

newGetDataEndpoint

Arguments

:: APIName
$sel:aPIName:GetDataEndpoint'

-> GetDataEndpoint

Create a value of GetDataEndpoint with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:streamARN:GetDataEndpoint', getDataEndpoint_streamARN - The Amazon Resource Name (ARN) of the stream that you want to get the endpoint for. You must specify either this parameter or a StreamName in the request.

$sel:streamName:GetDataEndpoint', getDataEndpoint_streamName - The name of the stream that you want to get the endpoint for. You must specify either this parameter or a StreamARN in the request.

$sel:aPIName:GetDataEndpoint', getDataEndpoint_aPIName - The name of the API action for which to get an endpoint.

Request Lenses

getDataEndpoint_streamARN :: Lens' GetDataEndpoint (Maybe Text)

The Amazon Resource Name (ARN) of the stream that you want to get the endpoint for. You must specify either this parameter or a StreamName in the request.

getDataEndpoint_streamName :: Lens' GetDataEndpoint (Maybe Text)

The name of the stream that you want to get the endpoint for. You must specify either this parameter or a StreamARN in the request.

getDataEndpoint_aPIName :: Lens' GetDataEndpoint APIName

The name of the API action for which to get an endpoint.

Destructuring the Response

data GetDataEndpointResponse

See: newGetDataEndpointResponse smart constructor.

Constructors

GetDataEndpointResponse'

Fields

dataEndpoint :: Maybe Text

The endpoint value. To read data from the stream or to write data to it, specify this endpoint in your application.

httpStatus :: Int

The response's http status code.

newGetDataEndpointResponse

Arguments

:: Int
$sel:httpStatus:GetDataEndpointResponse'

-> GetDataEndpointResponse

Create a value of GetDataEndpointResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:dataEndpoint:GetDataEndpointResponse', getDataEndpointResponse_dataEndpoint - The endpoint value. To read data from the stream or to write data to it, specify this endpoint in your application.

$sel:httpStatus:GetDataEndpointResponse', getDataEndpointResponse_httpStatus - The response's http status code.

Response Lenses

getDataEndpointResponse_dataEndpoint :: Lens' GetDataEndpointResponse (Maybe Text)

The endpoint value. To read data from the stream or to write data to it, specify this endpoint in your application.

getDataEndpointResponse_httpStatus :: Lens' GetDataEndpointResponse Int

The response's http status code.