Amazonka.KinesisVideo.Types.ChannelInfo

Description

Documentation

data ChannelInfo

A structure that encapsulates a signaling channel's metadata and properties.

See: newChannelInfo smart constructor.

Constructors

ChannelInfo'

Fields

creationTime :: Maybe POSIX

The time at which the signaling channel was created.

channelStatus :: Maybe StreamStatus

Current status of the signaling channel.

channelARN :: Maybe Text

The Amazon Resource Name (ARN) of the signaling channel.

singleMasterConfiguration :: Maybe SingleMasterConfiguration

A structure that contains the configuration for the SINGLE_MASTER channel type.

channelName :: Maybe Text

The name of the signaling channel.

version :: Maybe Text

The current version of the signaling channel.

channelType :: Maybe ChannelType

The type of the signaling channel.

newChannelInfo :: ChannelInfo

Create a value of ChannelInfo with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:creationTime:ChannelInfo', channelInfo_creationTime - The time at which the signaling channel was created.

$sel:channelStatus:ChannelInfo', channelInfo_channelStatus - Current status of the signaling channel.

$sel:channelARN:ChannelInfo', channelInfo_channelARN - The Amazon Resource Name (ARN) of the signaling channel.

$sel:singleMasterConfiguration:ChannelInfo', channelInfo_singleMasterConfiguration - A structure that contains the configuration for the SINGLE_MASTER channel type.

$sel:channelName:ChannelInfo', channelInfo_channelName - The name of the signaling channel.

$sel:version:ChannelInfo', channelInfo_version - The current version of the signaling channel.

$sel:channelType:ChannelInfo', channelInfo_channelType - The type of the signaling channel.

channelInfo_creationTime :: Lens' ChannelInfo (Maybe UTCTime)

The time at which the signaling channel was created.

channelInfo_channelStatus :: Lens' ChannelInfo (Maybe StreamStatus)

Current status of the signaling channel.

channelInfo_channelARN :: Lens' ChannelInfo (Maybe Text)

The Amazon Resource Name (ARN) of the signaling channel.

channelInfo_singleMasterConfiguration :: Lens' ChannelInfo (Maybe SingleMasterConfiguration)

A structure that contains the configuration for the SINGLE_MASTER channel type.

channelInfo_channelName :: Lens' ChannelInfo (Maybe Text)

The name of the signaling channel.

channelInfo_version :: Lens' ChannelInfo (Maybe Text)

The current version of the signaling channel.

channelInfo_channelType :: Lens' ChannelInfo (Maybe ChannelType)

The type of the signaling channel.