Amazonka.KinesisVideo.Types.SingleMasterChannelEndpointConfiguration

Description

Documentation

data SingleMasterChannelEndpointConfiguration

An object that contains the endpoint configuration for the SINGLE_MASTER channel type.

See: newSingleMasterChannelEndpointConfiguration smart constructor.

Constructors

SingleMasterChannelEndpointConfiguration'

Fields

protocols :: Maybe (NonEmpty ChannelProtocol)

This property is used to determine the nature of communication over this SINGLE_MASTER signaling channel. If WSS is specified, this API returns a websocket endpoint. If HTTPS is specified, this API returns an HTTPS endpoint.

role' :: Maybe ChannelRole

This property is used to determine messaging permissions in this SINGLE_MASTER signaling channel. If MASTER is specified, this API returns an endpoint that a client can use to receive offers from and send answers to any of the viewers on this signaling channel. If VIEWER is specified, this API returns an endpoint that a client can use only to send offers to another MASTER client on this signaling channel.

newSingleMasterChannelEndpointConfiguration :: SingleMasterChannelEndpointConfiguration

Create a value of SingleMasterChannelEndpointConfiguration with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:protocols:SingleMasterChannelEndpointConfiguration', singleMasterChannelEndpointConfiguration_protocols - This property is used to determine the nature of communication over this SINGLE_MASTER signaling channel. If WSS is specified, this API returns a websocket endpoint. If HTTPS is specified, this API returns an HTTPS endpoint.

$sel:role':SingleMasterChannelEndpointConfiguration', singleMasterChannelEndpointConfiguration_role - This property is used to determine messaging permissions in this SINGLE_MASTER signaling channel. If MASTER is specified, this API returns an endpoint that a client can use to receive offers from and send answers to any of the viewers on this signaling channel. If VIEWER is specified, this API returns an endpoint that a client can use only to send offers to another MASTER client on this signaling channel.

singleMasterChannelEndpointConfiguration_protocols :: Lens' SingleMasterChannelEndpointConfiguration (Maybe (NonEmpty ChannelProtocol))

This property is used to determine the nature of communication over this SINGLE_MASTER signaling channel. If WSS is specified, this API returns a websocket endpoint. If HTTPS is specified, this API returns an HTTPS endpoint.

singleMasterChannelEndpointConfiguration_role :: Lens' SingleMasterChannelEndpointConfiguration (Maybe ChannelRole)

This property is used to determine messaging permissions in this SINGLE_MASTER signaling channel. If MASTER is specified, this API returns an endpoint that a client can use to receive offers from and send answers to any of the viewers on this signaling channel. If VIEWER is specified, this API returns an endpoint that a client can use only to send offers to another MASTER client on this signaling channel.