Amazonka.KinesisVideoArchivedMedia.GetClip

Description

Downloads an MP4 file (clip) containing the archived, on-demand media from the specified video stream over the specified time range.

Both the StreamName and the StreamARN parameters are optional, but you must specify either the StreamName or the StreamARN when invoking this API operation.

As a prerequisite to using GetCLip API, you must obtain an endpoint using GetDataEndpoint, specifying GET_CLIP for@ the APIName@ parameter.

An Amazon Kinesis video stream has the following requirements for providing data through MP4:

The media must contain h.264 or h.265 encoded video and, optionally, AAC or G.711 encoded audio. Specifically, the codec ID of track 1 should be V_MPEG/ISO/AVC (for h.264) or V_MPEGH/ISO/HEVC (for H.265). Optionally, the codec ID of track 2 should be A_AAC (for AAC) or A_MS/ACM (for G.711).

Data retention must be greater than 0.

The video track of each fragment must contain codec private data in the Advanced Video Coding (AVC) for H.264 format and HEVC for H.265 format. For more information, see MPEG-4 specification ISO/IEC 14496-15. For information about adapting stream data to a given format, see NAL Adaptation Flags.

The audio track (if present) of each fragment must contain codec private data in the AAC format (AAC specification ISO/IEC 13818-7) or the MS Wave format.

You can monitor the amount of outgoing data by monitoring the GetClip.OutgoingBytes Amazon CloudWatch metric. For information about using CloudWatch to monitor Kinesis Video Streams, see Monitoring Kinesis Video Streams. For pricing information, see Amazon Kinesis Video Streams Pricing and AWS Pricing. Charges for outgoing AWS data apply.

Creating a Request

data GetClip

See: newGetClip smart constructor.

Constructors

GetClip'

Fields

streamARN :: Maybe Text

The Amazon Resource Name (ARN) of the stream for which to retrieve the media clip.

You must specify either the StreamName or the StreamARN.

streamName :: Maybe Text

The name of the stream for which to retrieve the media clip.

You must specify either the StreamName or the StreamARN.

clipFragmentSelector :: ClipFragmentSelector

The time range of the requested clip and the source of the timestamps.

newGetClip

Arguments

:: ClipFragmentSelector
$sel:clipFragmentSelector:GetClip'

-> GetClip

Create a value of GetClip with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:streamARN:GetClip', getClip_streamARN - The Amazon Resource Name (ARN) of the stream for which to retrieve the media clip.

You must specify either the StreamName or the StreamARN.

$sel:streamName:GetClip', getClip_streamName - The name of the stream for which to retrieve the media clip.

You must specify either the StreamName or the StreamARN.

$sel:clipFragmentSelector:GetClip', getClip_clipFragmentSelector - The time range of the requested clip and the source of the timestamps.

Request Lenses

getClip_streamARN :: Lens' GetClip (Maybe Text)

The Amazon Resource Name (ARN) of the stream for which to retrieve the media clip.

You must specify either the StreamName or the StreamARN.

getClip_streamName :: Lens' GetClip (Maybe Text)

The name of the stream for which to retrieve the media clip.

You must specify either the StreamName or the StreamARN.

getClip_clipFragmentSelector :: Lens' GetClip ClipFragmentSelector

The time range of the requested clip and the source of the timestamps.

Destructuring the Response

data GetClipResponse

See: newGetClipResponse smart constructor.

Constructors

GetClipResponse'

Fields

contentType :: Maybe Text

The content type of the media in the requested clip.

httpStatus :: Int

The response's http status code.

payload :: ResponseBody

Traditional MP4 file that contains the media clip from the specified video stream. The output will contain the first 100 MB or the first 200 fragments from the specified start timestamp. For more information, see Kinesis Video Streams Limits.

newGetClipResponse

Arguments

:: Int
$sel:httpStatus:GetClipResponse'

-> ResponseBody
$sel:payload:GetClipResponse'

-> GetClipResponse

Create a value of GetClipResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:contentType:GetClipResponse', getClipResponse_contentType - The content type of the media in the requested clip.

$sel:httpStatus:GetClipResponse', getClipResponse_httpStatus - The response's http status code.

$sel:payload:GetClipResponse', getClipResponse_payload - Traditional MP4 file that contains the media clip from the specified video stream. The output will contain the first 100 MB or the first 200 fragments from the specified start timestamp. For more information, see Kinesis Video Streams Limits.

Response Lenses

getClipResponse_contentType :: Lens' GetClipResponse (Maybe Text)

The content type of the media in the requested clip.

getClipResponse_httpStatus :: Lens' GetClipResponse Int

The response's http status code.

getClipResponse_payload :: Lens' GetClipResponse ResponseBody

Traditional MP4 file that contains the media clip from the specified video stream. The output will contain the first 100 MB or the first 200 fragments from the specified start timestamp. For more information, see Kinesis Video Streams Limits.