Amazonka.KinesisVideoArchivedMedia.Types.FragmentSelector

Description

Documentation

data FragmentSelector

Describes the timestamp range and timestamp origin of a range of fragments.

Only fragments with a start timestamp greater than or equal to the given start time and less than or equal to the end time are returned. For example, if a stream contains fragments with the following start timestamps:

00:00:00

00:00:02

00:00:04

00:00:06

A fragment selector range with a start time of 00:00:01 and end time of 00:00:04 would return the fragments with start times of 00:00:02 and 00:00:04.

See: newFragmentSelector smart constructor.

Constructors

FragmentSelector'

Fields

fragmentSelectorType :: FragmentSelectorType

The origin of the timestamps to use (Server or Producer).

timestampRange :: TimestampRange

The range of timestamps to return.

newFragmentSelector

Arguments

:: FragmentSelectorType
$sel:fragmentSelectorType:FragmentSelector'

-> TimestampRange
$sel:timestampRange:FragmentSelector'

-> FragmentSelector

Create a value of FragmentSelector with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:fragmentSelectorType:FragmentSelector', fragmentSelector_fragmentSelectorType - The origin of the timestamps to use (Server or Producer).

$sel:timestampRange:FragmentSelector', fragmentSelector_timestampRange - The range of timestamps to return.

fragmentSelector_fragmentSelectorType :: Lens' FragmentSelector FragmentSelectorType

The origin of the timestamps to use (Server or Producer).

fragmentSelector_timestampRange :: Lens' FragmentSelector TimestampRange

The range of timestamps to return.