Amazonka.KinesisVideoSignaling

Description

Derived from API version 2019-12-04 of the AWS service descriptions, licensed under Apache 2.0.

Kinesis Video Streams Signaling Service is a intermediate service that establishes a communication channel for discovering peers, transmitting offers and answers in order to establish peer-to-peer connection in webRTC technology.

Service Configuration

defaultService :: Service

API version 2019-12-04 of the Amazon Kinesis Video Signaling Channels SDK configuration.

Errors

Error matchers are designed for use with the functions provided by Control.Exception.Lens. This allows catching (and rethrowing) service specific errors returned by KinesisVideoSignaling.

InvalidArgumentException

_InvalidArgumentException :: AsError a => Getting (First ServiceError) a ServiceError

The value for this input parameter is invalid.

NotAuthorizedException

_NotAuthorizedException :: AsError a => Getting (First ServiceError) a ServiceError

The caller is not authorized to perform this operation.

ClientLimitExceededException

_ClientLimitExceededException :: AsError a => Getting (First ServiceError) a ServiceError

Your request was throttled because you have exceeded the limit of allowed client calls. Try making the call later.

SessionExpiredException

_SessionExpiredException :: AsError a => Getting (First ServiceError) a ServiceError

If the client session is expired. Once the client is connected, the session is valid for 45 minutes. Client should reconnect to the channel to continue sending/receiving messages.

InvalidClientException

_InvalidClientException :: AsError a => Getting (First ServiceError) a ServiceError

The specified client is invalid.

ResourceNotFoundException

_ResourceNotFoundException :: AsError a => Getting (First ServiceError) a ServiceError

The specified resource is not found.

Waiters

Waiters poll by repeatedly sending a request until some remote success condition configured by the Wait specification is fulfilled. The Wait specification determines how many attempts should be made, in addition to delay and retry strategies.

Operations

Some AWS operations return results that are incomplete and require subsequent requests in order to obtain the entire result set. The process of sending subsequent requests to continue where a previous request left off is called pagination. For example, the ListObjects operation of Amazon S3 returns up to 1000 objects at a time, and you must send subsequent requests with the appropriate Marker in order to retrieve the next page of results.

Operations that have an AWSPager instance can transparently perform subsequent requests, correctly setting Markers and other request facets to iterate through the entire result set of a truncated API operation. Operations which support this have an additional note in the documentation.

Many operations have the ability to filter results on the server side. See the individual operation parameters for details.

SendAlexaOfferToMaster

data SendAlexaOfferToMaster

See: newSendAlexaOfferToMaster smart constructor.

Constructors

SendAlexaOfferToMaster' Text Text Text

newSendAlexaOfferToMaster

Arguments

:: Text
$sel:channelARN:SendAlexaOfferToMaster'

-> Text
$sel:senderClientId:SendAlexaOfferToMaster'

-> Text
$sel:messagePayload:SendAlexaOfferToMaster'

-> SendAlexaOfferToMaster

Create a value of SendAlexaOfferToMaster with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:channelARN:SendAlexaOfferToMaster', sendAlexaOfferToMaster_channelARN - The ARN of the signaling channel by which Alexa and the master peer communicate.

$sel:senderClientId:SendAlexaOfferToMaster', sendAlexaOfferToMaster_senderClientId - The unique identifier for the sender client.

$sel:messagePayload:SendAlexaOfferToMaster', sendAlexaOfferToMaster_messagePayload - The base64-encoded SDP offer content.

data SendAlexaOfferToMasterResponse

See: newSendAlexaOfferToMasterResponse smart constructor.

Constructors

SendAlexaOfferToMasterResponse' (Maybe Text) Int

newSendAlexaOfferToMasterResponse

Arguments

:: Int
$sel:httpStatus:SendAlexaOfferToMasterResponse'

-> SendAlexaOfferToMasterResponse

Create a value of SendAlexaOfferToMasterResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:answer:SendAlexaOfferToMasterResponse', sendAlexaOfferToMasterResponse_answer - The base64-encoded SDP answer content.

$sel:httpStatus:SendAlexaOfferToMasterResponse', sendAlexaOfferToMasterResponse_httpStatus - The response's http status code.

GetIceServerConfig

data GetIceServerConfig

See: newGetIceServerConfig smart constructor.

Constructors

GetIceServerConfig' (Maybe Text) (Maybe Service) (Maybe Text) Text

newGetIceServerConfig

Arguments

:: Text
$sel:channelARN:GetIceServerConfig'

-> GetIceServerConfig

Create a value of GetIceServerConfig with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:clientId:GetIceServerConfig', getIceServerConfig_clientId - Unique identifier for the viewer. Must be unique within the signaling channel.

$sel:service:GetIceServerConfig', getIceServerConfig_service - Specifies the desired service. Currently, TURN is the only valid value.

$sel:username:GetIceServerConfig', getIceServerConfig_username - An optional user ID to be associated with the credentials.

$sel:channelARN:GetIceServerConfig', getIceServerConfig_channelARN - The ARN of the signaling channel to be used for the peer-to-peer connection between configured peers.

data GetIceServerConfigResponse

See: newGetIceServerConfigResponse smart constructor.

Constructors

GetIceServerConfigResponse' (Maybe [IceServer]) Int

newGetIceServerConfigResponse

Arguments

:: Int
$sel:httpStatus:GetIceServerConfigResponse'

-> GetIceServerConfigResponse

Create a value of GetIceServerConfigResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:iceServerList:GetIceServerConfigResponse', getIceServerConfigResponse_iceServerList - The list of ICE server information objects.

$sel:httpStatus:GetIceServerConfigResponse', getIceServerConfigResponse_httpStatus - The response's http status code.

Types

Service

newtype Service

Constructors

Service'

Fields

fromService :: Text

Bundled Patterns

pattern Service_TURN :: Service

IceServer

data IceServer

A structure for the ICE server connection data.

See: newIceServer smart constructor.

Constructors

IceServer' (Maybe Natural) (Maybe [Text]) (Maybe Text) (Maybe Text)

newIceServer :: IceServer

Create a value of IceServer with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:ttl:IceServer', iceServer_ttl - The period of time, in seconds, during which the username and password are valid.

$sel:uris:IceServer', iceServer_uris - An array of URIs, in the form specified in the I-D.petithuguenin-behave-turn-uris spec. These URIs provide the different addresses and/or protocols that can be used to reach the TURN server.

$sel:username:IceServer', iceServer_username - A username to login to the ICE server.

$sel:password:IceServer', iceServer_password - A password to login to the ICE server.