Amazonka.KinesisVideoSignaling.Types

Description

Service Configuration

defaultService :: Service

API version 2019-12-04 of the Amazon Kinesis Video Signaling Channels SDK configuration.

Errors

_InvalidArgumentException :: AsError a => Getting (First ServiceError) a ServiceError

The value for this input parameter is invalid.

_NotAuthorizedException :: AsError a => Getting (First ServiceError) a ServiceError

The caller is not authorized to perform this operation.

_ClientLimitExceededException :: AsError a => Getting (First ServiceError) a ServiceError

Your request was throttled because you have exceeded the limit of allowed client calls. Try making the call later.

_SessionExpiredException :: AsError a => Getting (First ServiceError) a ServiceError

If the client session is expired. Once the client is connected, the session is valid for 45 minutes. Client should reconnect to the channel to continue sending/receiving messages.

_InvalidClientException :: AsError a => Getting (First ServiceError) a ServiceError

The specified client is invalid.

_ResourceNotFoundException :: AsError a => Getting (First ServiceError) a ServiceError

The specified resource is not found.

Service

newtype Service

Constructors

Service'

Fields

fromService :: Text

Bundled Patterns

pattern Service_TURN :: Service

IceServer

data IceServer

A structure for the ICE server connection data.

See: newIceServer smart constructor.

Constructors

IceServer'

Fields

ttl :: Maybe Natural

The period of time, in seconds, during which the username and password are valid.

uris :: Maybe [Text]

An array of URIs, in the form specified in the I-D.petithuguenin-behave-turn-uris spec. These URIs provide the different addresses and/or protocols that can be used to reach the TURN server.

username :: Maybe Text

A username to login to the ICE server.

password :: Maybe Text

A password to login to the ICE server.

newIceServer :: IceServer

Create a value of IceServer with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:ttl:IceServer', iceServer_ttl - The period of time, in seconds, during which the username and password are valid.

$sel:uris:IceServer', iceServer_uris - An array of URIs, in the form specified in the I-D.petithuguenin-behave-turn-uris spec. These URIs provide the different addresses and/or protocols that can be used to reach the TURN server.

$sel:username:IceServer', iceServer_username - A username to login to the ICE server.

$sel:password:IceServer', iceServer_password - A password to login to the ICE server.

iceServer_ttl :: Lens' IceServer (Maybe Natural)

The period of time, in seconds, during which the username and password are valid.

iceServer_uris :: Lens' IceServer (Maybe [Text])

An array of URIs, in the form specified in the I-D.petithuguenin-behave-turn-uris spec. These URIs provide the different addresses and/or protocols that can be used to reach the TURN server.

iceServer_username :: Lens' IceServer (Maybe Text)

A username to login to the ICE server.

iceServer_password :: Lens' IceServer (Maybe Text)

A password to login to the ICE server.