Amazonka.LakeFormation.DeleteLFTag

Description

Deletes the specified tag key name. If the attribute key does not exist or the tag does not exist, then the operation will not do anything. If the attribute key exists, then the operation checks if any resources are tagged with this attribute key, if yes, the API throws a 400 Exception with the message "Delete not allowed" as the tag key is still attached with resources. You can consider untagging resources with this tag key.

Creating a Request

data DeleteLFTag

See: newDeleteLFTag smart constructor.

Constructors

DeleteLFTag'

Fields

catalogId :: Maybe Text

The identifier for the Data Catalog. By default, the account ID. The Data Catalog is the persistent metadata store. It contains database definitions, table definitions, and other control information to manage your AWS Lake Formation environment.

tagKey :: Text

The key-name for the tag to delete.

newDeleteLFTag

Arguments

:: Text
$sel:tagKey:DeleteLFTag'

-> DeleteLFTag

Create a value of DeleteLFTag with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:catalogId:DeleteLFTag', deleteLFTag_catalogId - The identifier for the Data Catalog. By default, the account ID. The Data Catalog is the persistent metadata store. It contains database definitions, table definitions, and other control information to manage your AWS Lake Formation environment.

$sel:tagKey:DeleteLFTag', deleteLFTag_tagKey - The key-name for the tag to delete.

Request Lenses

deleteLFTag_catalogId :: Lens' DeleteLFTag (Maybe Text)

The identifier for the Data Catalog. By default, the account ID. The Data Catalog is the persistent metadata store. It contains database definitions, table definitions, and other control information to manage your AWS Lake Formation environment.

deleteLFTag_tagKey :: Lens' DeleteLFTag Text

The key-name for the tag to delete.

Destructuring the Response

data DeleteLFTagResponse

See: newDeleteLFTagResponse smart constructor.

Constructors

DeleteLFTagResponse'

Fields

httpStatus :: Int

The response's http status code.

newDeleteLFTagResponse

Arguments

:: Int
$sel:httpStatus:DeleteLFTagResponse'

-> DeleteLFTagResponse

Create a value of DeleteLFTagResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:httpStatus:DeleteLFTagResponse', deleteLFTagResponse_httpStatus - The response's http status code.

Response Lenses

deleteLFTagResponse_httpStatus :: Lens' DeleteLFTagResponse Int

The response's http status code.