Amazonka.LakeFormation.Lens

Description

Operations

BatchRevokePermissions

batchRevokePermissions_catalogId :: Lens' BatchRevokePermissions (Maybe Text)

The identifier for the Data Catalog. By default, the account ID. The Data Catalog is the persistent metadata store. It contains database definitions, table definitions, and other control information to manage your AWS Lake Formation environment.

batchRevokePermissions_entries :: Lens' BatchRevokePermissions [BatchPermissionsRequestEntry]

A list of up to 20 entries for resource permissions to be revoked by batch operation to the principal.

batchRevokePermissionsResponse_failures :: Lens' BatchRevokePermissionsResponse (Maybe [BatchPermissionsFailureEntry])

A list of failures to revoke permissions to the resources.

batchRevokePermissionsResponse_httpStatus :: Lens' BatchRevokePermissionsResponse Int

The response's http status code.

DescribeResource

describeResource_resourceArn :: Lens' DescribeResource Text

The resource ARN.

describeResourceResponse_resourceInfo :: Lens' DescribeResourceResponse (Maybe ResourceInfo)

A structure containing information about an AWS Lake Formation resource.

describeResourceResponse_httpStatus :: Lens' DescribeResourceResponse Int

The response's http status code.

BatchGrantPermissions

batchGrantPermissions_catalogId :: Lens' BatchGrantPermissions (Maybe Text)

The identifier for the Data Catalog. By default, the account ID. The Data Catalog is the persistent metadata store. It contains database definitions, table definitions, and other control information to manage your AWS Lake Formation environment.

batchGrantPermissions_entries :: Lens' BatchGrantPermissions [BatchPermissionsRequestEntry]

A list of up to 20 entries for resource permissions to be granted by batch operation to the principal.

batchGrantPermissionsResponse_failures :: Lens' BatchGrantPermissionsResponse (Maybe [BatchPermissionsFailureEntry])

A list of failures to grant permissions to the resources.

batchGrantPermissionsResponse_httpStatus :: Lens' BatchGrantPermissionsResponse Int

The response's http status code.

GetEffectivePermissionsForPath

getEffectivePermissionsForPath_catalogId :: Lens' GetEffectivePermissionsForPath (Maybe Text)

The identifier for the Data Catalog. By default, the account ID. The Data Catalog is the persistent metadata store. It contains database definitions, table definitions, and other control information to manage your AWS Lake Formation environment.

getEffectivePermissionsForPath_nextToken :: Lens' GetEffectivePermissionsForPath (Maybe Text)

A continuation token, if this is not the first call to retrieve this list.

getEffectivePermissionsForPath_maxResults :: Lens' GetEffectivePermissionsForPath (Maybe Natural)

The maximum number of results to return.

getEffectivePermissionsForPath_resourceArn :: Lens' GetEffectivePermissionsForPath Text

The Amazon Resource Name (ARN) of the resource for which you want to get permissions.

getEffectivePermissionsForPathResponse_nextToken :: Lens' GetEffectivePermissionsForPathResponse (Maybe Text)

A continuation token, if this is not the first call to retrieve this list.

getEffectivePermissionsForPathResponse_permissions :: Lens' GetEffectivePermissionsForPathResponse (Maybe [PrincipalResourcePermissions])

A list of the permissions for the specified table or database resource located at the path in Amazon S3.

getEffectivePermissionsForPathResponse_httpStatus :: Lens' GetEffectivePermissionsForPathResponse Int

The response's http status code.

RevokePermissions

revokePermissions_catalogId :: Lens' RevokePermissions (Maybe Text)

The identifier for the Data Catalog. By default, the account ID. The Data Catalog is the persistent metadata store. It contains database definitions, table definitions, and other control information to manage your AWS Lake Formation environment.

revokePermissions_permissionsWithGrantOption :: Lens' RevokePermissions (Maybe [Permission])

Indicates a list of permissions for which to revoke the grant option allowing the principal to pass permissions to other principals.

revokePermissions_principal :: Lens' RevokePermissions DataLakePrincipal

The principal to be revoked permissions on the resource.

revokePermissions_resource :: Lens' RevokePermissions Resource

The resource to which permissions are to be revoked.

revokePermissions_permissions :: Lens' RevokePermissions [Permission]

The permissions revoked to the principal on the resource. For information about permissions, see Security and Access Control to Metadata and Data.

revokePermissionsResponse_httpStatus :: Lens' RevokePermissionsResponse Int

The response's http status code.

UpdateResource

updateResource_roleArn :: Lens' UpdateResource Text

The new role to use for the given resource registered in AWS Lake Formation.

updateResource_resourceArn :: Lens' UpdateResource Text

The resource ARN.

updateResourceResponse_httpStatus :: Lens' UpdateResourceResponse Int

The response's http status code.

AddLFTagsToResource

addLFTagsToResource_catalogId :: Lens' AddLFTagsToResource (Maybe Text)

The identifier for the Data Catalog. By default, the account ID. The Data Catalog is the persistent metadata store. It contains database definitions, table definitions, and other control information to manage your AWS Lake Formation environment.

addLFTagsToResource_resource :: Lens' AddLFTagsToResource Resource

The resource to which to attach a tag.

addLFTagsToResource_lFTags :: Lens' AddLFTagsToResource (NonEmpty LFTagPair)

The tags to attach to the resource.

addLFTagsToResourceResponse_failures :: Lens' AddLFTagsToResourceResponse (Maybe [LFTagError])

A list of failures to tag the resource.

addLFTagsToResourceResponse_httpStatus :: Lens' AddLFTagsToResourceResponse Int

The response's http status code.

SearchTablesByLFTags

searchTablesByLFTags_catalogId :: Lens' SearchTablesByLFTags (Maybe Text)

The identifier for the Data Catalog. By default, the account ID. The Data Catalog is the persistent metadata store. It contains database definitions, table definitions, and other control information to manage your AWS Lake Formation environment.

searchTablesByLFTags_nextToken :: Lens' SearchTablesByLFTags (Maybe Text)

A continuation token, if this is not the first call to retrieve this list.

searchTablesByLFTags_maxResults :: Lens' SearchTablesByLFTags (Maybe Natural)

The maximum number of results to return.

searchTablesByLFTags_expression :: Lens' SearchTablesByLFTags (NonEmpty LFTag)

A list of conditions (LFTag structures) to search for in table resources.

searchTablesByLFTagsResponse_tableList :: Lens' SearchTablesByLFTagsResponse (Maybe [TaggedTable])

A list of tables that meet the tag conditions.

searchTablesByLFTagsResponse_nextToken :: Lens' SearchTablesByLFTagsResponse (Maybe Text)

A continuation token, present if the current list segment is not the last.

searchTablesByLFTagsResponse_httpStatus :: Lens' SearchTablesByLFTagsResponse Int

The response's http status code.

ListResources

listResources_filterConditionList :: Lens' ListResources (Maybe (NonEmpty FilterCondition))

Any applicable row-level and/or column-level filtering conditions for the resources.

listResources_nextToken :: Lens' ListResources (Maybe Text)

A continuation token, if this is not the first call to retrieve these resources.

listResources_maxResults :: Lens' ListResources (Maybe Natural)

The maximum number of resource results.

listResourcesResponse_resourceInfoList :: Lens' ListResourcesResponse (Maybe [ResourceInfo])

A summary of the data lake resources.

listResourcesResponse_nextToken :: Lens' ListResourcesResponse (Maybe Text)

A continuation token, if this is not the first call to retrieve these resources.

listResourcesResponse_httpStatus :: Lens' ListResourcesResponse Int

The response's http status code.

GetLFTag

getLFTag_catalogId :: Lens' GetLFTag (Maybe Text)

The identifier for the Data Catalog. By default, the account ID. The Data Catalog is the persistent metadata store. It contains database definitions, table definitions, and other control information to manage your AWS Lake Formation environment.

getLFTag_tagKey :: Lens' GetLFTag Text

The key-name for the tag.

getLFTagResponse_tagValues :: Lens' GetLFTagResponse (Maybe (NonEmpty Text))

A list of possible values an attribute can take.

getLFTagResponse_catalogId :: Lens' GetLFTagResponse (Maybe Text)

The identifier for the Data Catalog. By default, the account ID. The Data Catalog is the persistent metadata store. It contains database definitions, table definitions, and other control information to manage your AWS Lake Formation environment.

getLFTagResponse_tagKey :: Lens' GetLFTagResponse (Maybe Text)

The key-name for the tag.

getLFTagResponse_httpStatus :: Lens' GetLFTagResponse Int

The response's http status code.

RemoveLFTagsFromResource

removeLFTagsFromResource_catalogId :: Lens' RemoveLFTagsFromResource (Maybe Text)

The identifier for the Data Catalog. By default, the account ID. The Data Catalog is the persistent metadata store. It contains database definitions, table definitions, and other control information to manage your AWS Lake Formation environment.

removeLFTagsFromResource_resource :: Lens' RemoveLFTagsFromResource Resource

The resource where you want to remove a tag.

removeLFTagsFromResource_lFTags :: Lens' RemoveLFTagsFromResource (NonEmpty LFTagPair)

The tags to be removed from the resource.

removeLFTagsFromResourceResponse_failures :: Lens' RemoveLFTagsFromResourceResponse (Maybe [LFTagError])

A list of failures to untag a resource.

removeLFTagsFromResourceResponse_httpStatus :: Lens' RemoveLFTagsFromResourceResponse Int

The response's http status code.

UpdateLFTag

updateLFTag_catalogId :: Lens' UpdateLFTag (Maybe Text)

The identifier for the Data Catalog. By default, the account ID. The Data Catalog is the persistent metadata store. It contains database definitions, table definitions, and other control information to manage your AWS Lake Formation environment.

updateLFTag_tagValuesToAdd :: Lens' UpdateLFTag (Maybe (NonEmpty Text))

A list of tag values to add from the tag.

updateLFTag_tagValuesToDelete :: Lens' UpdateLFTag (Maybe (NonEmpty Text))

A list of tag values to delete from the tag.

updateLFTag_tagKey :: Lens' UpdateLFTag Text

The key-name for the tag for which to add or delete values.

updateLFTagResponse_httpStatus :: Lens' UpdateLFTagResponse Int

The response's http status code.

DeleteLFTag

deleteLFTag_catalogId :: Lens' DeleteLFTag (Maybe Text)

The identifier for the Data Catalog. By default, the account ID. The Data Catalog is the persistent metadata store. It contains database definitions, table definitions, and other control information to manage your AWS Lake Formation environment.

deleteLFTag_tagKey :: Lens' DeleteLFTag Text

The key-name for the tag to delete.

deleteLFTagResponse_httpStatus :: Lens' DeleteLFTagResponse Int

The response's http status code.

CreateLFTag

createLFTag_catalogId :: Lens' CreateLFTag (Maybe Text)

The identifier for the Data Catalog. By default, the account ID. The Data Catalog is the persistent metadata store. It contains database definitions, table definitions, and other control information to manage your AWS Lake Formation environment.

createLFTag_tagKey :: Lens' CreateLFTag Text

The key-name for the tag.

createLFTag_tagValues :: Lens' CreateLFTag (NonEmpty Text)

A list of possible values an attribute can take.

createLFTagResponse_httpStatus :: Lens' CreateLFTagResponse Int

The response's http status code.

GetResourceLFTags

getResourceLFTags_showAssignedLFTags :: Lens' GetResourceLFTags (Maybe Bool)

Indicates whether to show the assigned tags.

getResourceLFTags_catalogId :: Lens' GetResourceLFTags (Maybe Text)

The identifier for the Data Catalog. By default, the account ID. The Data Catalog is the persistent metadata store. It contains database definitions, table definitions, and other control information to manage your AWS Lake Formation environment.

getResourceLFTags_resource :: Lens' GetResourceLFTags Resource

The resource for which you want to return tags.

getResourceLFTagsResponse_lFTagsOnTable :: Lens' GetResourceLFTagsResponse (Maybe (NonEmpty LFTagPair))

A list of tags applied to a table resource.

getResourceLFTagsResponse_lFTagOnDatabase :: Lens' GetResourceLFTagsResponse (Maybe (NonEmpty LFTagPair))

A list of tags applied to a database resource.

getResourceLFTagsResponse_lFTagsOnColumns :: Lens' GetResourceLFTagsResponse (Maybe [ColumnLFTag])

A list of tags applied to a column resource.

getResourceLFTagsResponse_httpStatus :: Lens' GetResourceLFTagsResponse Int

The response's http status code.

PutDataLakeSettings

putDataLakeSettings_catalogId :: Lens' PutDataLakeSettings (Maybe Text)

The identifier for the Data Catalog. By default, the account ID. The Data Catalog is the persistent metadata store. It contains database definitions, table definitions, and other control information to manage your AWS Lake Formation environment.

putDataLakeSettings_dataLakeSettings :: Lens' PutDataLakeSettings DataLakeSettings

A structure representing a list of AWS Lake Formation principals designated as data lake administrators.

putDataLakeSettingsResponse_httpStatus :: Lens' PutDataLakeSettingsResponse Int

The response's http status code.

ListPermissions

listPermissions_resourceType :: Lens' ListPermissions (Maybe DataLakeResourceType)

Specifies a resource type to filter the permissions returned.

listPermissions_catalogId :: Lens' ListPermissions (Maybe Text)

The identifier for the Data Catalog. By default, the account ID. The Data Catalog is the persistent metadata store. It contains database definitions, table definitions, and other control information to manage your AWS Lake Formation environment.

listPermissions_nextToken :: Lens' ListPermissions (Maybe Text)

A continuation token, if this is not the first call to retrieve this list.

listPermissions_principal :: Lens' ListPermissions (Maybe DataLakePrincipal)

Specifies a principal to filter the permissions returned.

listPermissions_resource :: Lens' ListPermissions (Maybe Resource)

A resource where you will get a list of the principal permissions.

This operation does not support getting privileges on a table with columns. Instead, call this operation on the table, and the operation returns the table and the table w columns.

listPermissions_maxResults :: Lens' ListPermissions (Maybe Natural)

The maximum number of results to return.

listPermissionsResponse_nextToken :: Lens' ListPermissionsResponse (Maybe Text)

A continuation token, if this is not the first call to retrieve this list.

listPermissionsResponse_principalResourcePermissions :: Lens' ListPermissionsResponse (Maybe [PrincipalResourcePermissions])

A list of principals and their permissions on the resource for the specified principal and resource types.

listPermissionsResponse_httpStatus :: Lens' ListPermissionsResponse Int

The response's http status code.

DeregisterResource

deregisterResource_resourceArn :: Lens' DeregisterResource Text

The Amazon Resource Name (ARN) of the resource that you want to deregister.

deregisterResourceResponse_httpStatus :: Lens' DeregisterResourceResponse Int

The response's http status code.

GetDataLakeSettings

getDataLakeSettings_catalogId :: Lens' GetDataLakeSettings (Maybe Text)

The identifier for the Data Catalog. By default, the account ID. The Data Catalog is the persistent metadata store. It contains database definitions, table definitions, and other control information to manage your AWS Lake Formation environment.

getDataLakeSettingsResponse_dataLakeSettings :: Lens' GetDataLakeSettingsResponse (Maybe DataLakeSettings)

A structure representing a list of AWS Lake Formation principals designated as data lake administrators.

getDataLakeSettingsResponse_httpStatus :: Lens' GetDataLakeSettingsResponse Int

The response's http status code.

SearchDatabasesByLFTags

searchDatabasesByLFTags_catalogId :: Lens' SearchDatabasesByLFTags (Maybe Text)

The identifier for the Data Catalog. By default, the account ID. The Data Catalog is the persistent metadata store. It contains database definitions, table definitions, and other control information to manage your AWS Lake Formation environment.

searchDatabasesByLFTags_nextToken :: Lens' SearchDatabasesByLFTags (Maybe Text)

A continuation token, if this is not the first call to retrieve this list.

searchDatabasesByLFTags_maxResults :: Lens' SearchDatabasesByLFTags (Maybe Natural)

The maximum number of results to return.

searchDatabasesByLFTags_expression :: Lens' SearchDatabasesByLFTags (NonEmpty LFTag)

A list of conditions (LFTag structures) to search for in database resources.

searchDatabasesByLFTagsResponse_databaseList :: Lens' SearchDatabasesByLFTagsResponse (Maybe [TaggedDatabase])

A list of databases that meet the tag conditions.

searchDatabasesByLFTagsResponse_nextToken :: Lens' SearchDatabasesByLFTagsResponse (Maybe Text)

A continuation token, present if the current list segment is not the last.

searchDatabasesByLFTagsResponse_httpStatus :: Lens' SearchDatabasesByLFTagsResponse Int

The response's http status code.

RegisterResource

registerResource_useServiceLinkedRole :: Lens' RegisterResource (Maybe Bool)

Designates an AWS Identity and Access Management (IAM) service-linked role by registering this role with the Data Catalog. A service-linked role is a unique type of IAM role that is linked directly to Lake Formation.

For more information, see Using Service-Linked Roles for Lake Formation.

registerResource_roleArn :: Lens' RegisterResource (Maybe Text)

The identifier for the role that registers the resource.

registerResource_resourceArn :: Lens' RegisterResource Text

The Amazon Resource Name (ARN) of the resource that you want to register.

registerResourceResponse_httpStatus :: Lens' RegisterResourceResponse Int

The response's http status code.

GrantPermissions

grantPermissions_catalogId :: Lens' GrantPermissions (Maybe Text)

The identifier for the Data Catalog. By default, the account ID. The Data Catalog is the persistent metadata store. It contains database definitions, table definitions, and other control information to manage your AWS Lake Formation environment.

grantPermissions_permissionsWithGrantOption :: Lens' GrantPermissions (Maybe [Permission])

Indicates a list of the granted permissions that the principal may pass to other users. These permissions may only be a subset of the permissions granted in the Privileges.

grantPermissions_principal :: Lens' GrantPermissions DataLakePrincipal

The principal to be granted the permissions on the resource. Supported principals are IAM users or IAM roles, and they are defined by their principal type and their ARN.

Note that if you define a resource with a particular ARN, then later delete, and recreate a resource with that same ARN, the resource maintains the permissions already granted.

grantPermissions_resource :: Lens' GrantPermissions Resource

The resource to which permissions are to be granted. Resources in AWS Lake Formation are the Data Catalog, databases, and tables.

grantPermissions_permissions :: Lens' GrantPermissions [Permission]

The permissions granted to the principal on the resource. AWS Lake Formation defines privileges to grant and revoke access to metadata in the Data Catalog and data organized in underlying data storage such as Amazon S3. AWS Lake Formation requires that each principal be authorized to perform a specific task on AWS Lake Formation resources.

grantPermissionsResponse_httpStatus :: Lens' GrantPermissionsResponse Int

The response's http status code.

ListLFTags

listLFTags_resourceShareType :: Lens' ListLFTags (Maybe ResourceShareType)

If resource share type is ALL, returns both in-account tags and shared tags that the requester has permission to view. If resource share type is FOREIGN, returns all share tags that the requester can view. If no resource share type is passed, lists tags in the given catalog ID that the requester has permission to view.

listLFTags_catalogId :: Lens' ListLFTags (Maybe Text)

The identifier for the Data Catalog. By default, the account ID. The Data Catalog is the persistent metadata store. It contains database definitions, table definitions, and other control information to manage your AWS Lake Formation environment.

listLFTags_nextToken :: Lens' ListLFTags (Maybe Text)

A continuation token, if this is not the first call to retrieve this list.

listLFTags_maxResults :: Lens' ListLFTags (Maybe Natural)

The maximum number of results to return.

listLFTagsResponse_nextToken :: Lens' ListLFTagsResponse (Maybe Text)

A continuation token, present if the current list segment is not the last.

listLFTagsResponse_lFTags :: Lens' ListLFTagsResponse (Maybe (NonEmpty LFTagPair))

A list of tags that the requested has permission to view.

listLFTagsResponse_httpStatus :: Lens' ListLFTagsResponse Int

The response's http status code.

Types

BatchPermissionsFailureEntry

batchPermissionsFailureEntry_error :: Lens' BatchPermissionsFailureEntry (Maybe ErrorDetail)

An error message that applies to the failure of the entry.

batchPermissionsFailureEntry_requestEntry :: Lens' BatchPermissionsFailureEntry (Maybe BatchPermissionsRequestEntry)

An identifier for an entry of the batch request.

BatchPermissionsRequestEntry

batchPermissionsRequestEntry_permissionsWithGrantOption :: Lens' BatchPermissionsRequestEntry (Maybe [Permission])

Indicates if the option to pass permissions is granted.

batchPermissionsRequestEntry_principal :: Lens' BatchPermissionsRequestEntry (Maybe DataLakePrincipal)

The principal to be granted a permission.

batchPermissionsRequestEntry_resource :: Lens' BatchPermissionsRequestEntry (Maybe Resource)

The resource to which the principal is to be granted a permission.

batchPermissionsRequestEntry_permissions :: Lens' BatchPermissionsRequestEntry (Maybe [Permission])

The permissions to be granted.

batchPermissionsRequestEntry_id :: Lens' BatchPermissionsRequestEntry Text

A unique identifier for the batch permissions request entry.

CatalogResource

ColumnLFTag

columnLFTag_name :: Lens' ColumnLFTag (Maybe Text)

The name of a column resource.

columnLFTag_lFTags :: Lens' ColumnLFTag (Maybe (NonEmpty LFTagPair))

The tags attached to a column resource.

ColumnWildcard

columnWildcard_excludedColumnNames :: Lens' ColumnWildcard (Maybe [Text])

Excludes column names. Any column with this name will be excluded.

DataLakePrincipal

dataLakePrincipal_dataLakePrincipalIdentifier :: Lens' DataLakePrincipal (Maybe Text)

An identifier for the AWS Lake Formation principal.

DataLakeSettings

dataLakeSettings_dataLakeAdmins :: Lens' DataLakeSettings (Maybe [DataLakePrincipal])

A list of AWS Lake Formation principals. Supported principals are IAM users or IAM roles.

dataLakeSettings_trustedResourceOwners :: Lens' DataLakeSettings (Maybe [Text])

A list of the resource-owning account IDs that the caller's account can use to share their user access details (user ARNs). The user ARNs can be logged in the resource owner's AWS CloudTrail log.

You may want to specify this property when you are in a high-trust boundary, such as the same team or company.

dataLakeSettings_createDatabaseDefaultPermissions :: Lens' DataLakeSettings (Maybe [PrincipalPermissions])

A structure representing a list of up to three principal permissions entries for default create database permissions.

dataLakeSettings_createTableDefaultPermissions :: Lens' DataLakeSettings (Maybe [PrincipalPermissions])

A structure representing a list of up to three principal permissions entries for default create table permissions.

DataLocationResource

dataLocationResource_catalogId :: Lens' DataLocationResource (Maybe Text)

The identifier for the Data Catalog where the location is registered with AWS Lake Formation. By default, it is the account ID of the caller.

dataLocationResource_resourceArn :: Lens' DataLocationResource Text

The Amazon Resource Name (ARN) that uniquely identifies the data location resource.

DatabaseResource

databaseResource_catalogId :: Lens' DatabaseResource (Maybe Text)

The identifier for the Data Catalog. By default, it is the account ID of the caller.

databaseResource_name :: Lens' DatabaseResource Text

The name of the database resource. Unique to the Data Catalog.

DetailsMap

detailsMap_resourceShare :: Lens' DetailsMap (Maybe [Text])

A resource share ARN for a catalog resource shared through AWS Resource Access Manager (AWS RAM).

ErrorDetail

errorDetail_errorCode :: Lens' ErrorDetail (Maybe Text)

The code associated with this error.

errorDetail_errorMessage :: Lens' ErrorDetail (Maybe Text)

A message describing the error.

FilterCondition

filterCondition_field :: Lens' FilterCondition (Maybe FieldNameString)

The field to filter in the filter condition.

filterCondition_comparisonOperator :: Lens' FilterCondition (Maybe ComparisonOperator)

The comparison operator used in the filter condition.

filterCondition_stringValueList :: Lens' FilterCondition (Maybe [Text])

A string with values used in evaluating the filter condition.

LFTag

lFTag_tagKey :: Lens' LFTag Text

The key-name for the tag.

lFTag_tagValues :: Lens' LFTag (NonEmpty Text)

A list of possible values an attribute can take.

LFTagError

lFTagError_lFTag :: Lens' LFTagError (Maybe LFTagPair)

The key-name of the tag.

lFTagError_error :: Lens' LFTagError (Maybe ErrorDetail)

An error that occurred with the attachment or detachment of the tag.

LFTagKeyResource

lFTagKeyResource_catalogId :: Lens' LFTagKeyResource (Maybe Text)

The identifier for the Data Catalog. By default, the account ID. The Data Catalog is the persistent metadata store. It contains database definitions, table definitions, and other control information to manage your AWS Lake Formation environment.

lFTagKeyResource_tagKey :: Lens' LFTagKeyResource Text

The key-name for the tag.

lFTagKeyResource_tagValues :: Lens' LFTagKeyResource (NonEmpty Text)

A list of possible values an attribute can take.

LFTagPair

lFTagPair_catalogId :: Lens' LFTagPair (Maybe Text)

The identifier for the Data Catalog. By default, the account ID. The Data Catalog is the persistent metadata store. It contains database definitions, table definitions, and other control information to manage your AWS Lake Formation environment.

lFTagPair_tagKey :: Lens' LFTagPair Text

The key-name for the tag.

lFTagPair_tagValues :: Lens' LFTagPair (NonEmpty Text)

A list of possible values an attribute can take.

LFTagPolicyResource

lFTagPolicyResource_catalogId :: Lens' LFTagPolicyResource (Maybe Text)

The identifier for the Data Catalog. By default, the account ID. The Data Catalog is the persistent metadata store. It contains database definitions, table definitions, and other control information to manage your AWS Lake Formation environment.

lFTagPolicyResource_resourceType :: Lens' LFTagPolicyResource ResourceType

The resource type for which the tag policy applies.

lFTagPolicyResource_expression :: Lens' LFTagPolicyResource (NonEmpty LFTag)

A list of tag conditions that apply to the resource's tag policy.

PrincipalPermissions

principalPermissions_principal :: Lens' PrincipalPermissions (Maybe DataLakePrincipal)

The principal who is granted permissions.

principalPermissions_permissions :: Lens' PrincipalPermissions (Maybe [Permission])

The permissions that are granted to the principal.

PrincipalResourcePermissions

principalResourcePermissions_additionalDetails :: Lens' PrincipalResourcePermissions (Maybe DetailsMap)

This attribute can be used to return any additional details of PrincipalResourcePermissions. Currently returns only as a RAM resource share ARN.

principalResourcePermissions_permissionsWithGrantOption :: Lens' PrincipalResourcePermissions (Maybe [Permission])

Indicates whether to grant the ability to grant permissions (as a subset of permissions granted).

principalResourcePermissions_principal :: Lens' PrincipalResourcePermissions (Maybe DataLakePrincipal)

The Data Lake principal to be granted or revoked permissions.

principalResourcePermissions_resource :: Lens' PrincipalResourcePermissions (Maybe Resource)

The resource where permissions are to be granted or revoked.

principalResourcePermissions_permissions :: Lens' PrincipalResourcePermissions (Maybe [Permission])

The permissions to be granted or revoked on the resource.

Resource

resource_dataLocation :: Lens' Resource (Maybe DataLocationResource)

The location of an Amazon S3 path where permissions are granted or revoked.

resource_database :: Lens' Resource (Maybe DatabaseResource)

The database for the resource. Unique to the Data Catalog. A database is a set of associated table definitions organized into a logical group. You can Grant and Revoke database permissions to a principal.

resource_lFTag :: Lens' Resource (Maybe LFTagKeyResource)

The tag key and values attached to a resource.

resource_catalog :: Lens' Resource (Maybe CatalogResource)

The identifier for the Data Catalog. By default, the account ID. The Data Catalog is the persistent metadata store. It contains database definitions, table definitions, and other control information to manage your AWS Lake Formation environment.

resource_lFTagPolicy :: Lens' Resource (Maybe LFTagPolicyResource)

A list of tag conditions that define a resource's tag policy.

resource_table :: Lens' Resource (Maybe TableResource)

The table for the resource. A table is a metadata definition that represents your data. You can Grant and Revoke table privileges to a principal.

resource_tableWithColumns :: Lens' Resource (Maybe TableWithColumnsResource)

The table with columns for the resource. A principal with permissions to this resource can select metadata from the columns of a table in the Data Catalog and the underlying data in Amazon S3.

ResourceInfo

resourceInfo_resourceArn :: Lens' ResourceInfo (Maybe Text)

The Amazon Resource Name (ARN) of the resource.

resourceInfo_lastModified :: Lens' ResourceInfo (Maybe UTCTime)

The date and time the resource was last modified.

resourceInfo_roleArn :: Lens' ResourceInfo (Maybe Text)

The IAM role that registered a resource.

TableResource

tableResource_catalogId :: Lens' TableResource (Maybe Text)

The identifier for the Data Catalog. By default, it is the account ID of the caller.

tableResource_tableWildcard :: Lens' TableResource (Maybe TableWildcard)

A wildcard object representing every table under a database.

At least one of TableResource$Name or TableResource$TableWildcard is required.

tableResource_name :: Lens' TableResource (Maybe Text)

The name of the table.

tableResource_databaseName :: Lens' TableResource Text

The name of the database for the table. Unique to a Data Catalog. A database is a set of associated table definitions organized into a logical group. You can Grant and Revoke database privileges to a principal.

TableWildcard

TableWithColumnsResource

tableWithColumnsResource_catalogId :: Lens' TableWithColumnsResource (Maybe Text)

The identifier for the Data Catalog. By default, it is the account ID of the caller.

tableWithColumnsResource_columnWildcard :: Lens' TableWithColumnsResource (Maybe ColumnWildcard)

A wildcard specified by a ColumnWildcard object. At least one of ColumnNames or ColumnWildcard is required.

tableWithColumnsResource_columnNames :: Lens' TableWithColumnsResource (Maybe [Text])

The list of column names for the table. At least one of ColumnNames or ColumnWildcard is required.

tableWithColumnsResource_databaseName :: Lens' TableWithColumnsResource Text

The name of the database for the table with columns resource. Unique to the Data Catalog. A database is a set of associated table definitions organized into a logical group. You can Grant and Revoke database privileges to a principal.

tableWithColumnsResource_name :: Lens' TableWithColumnsResource Text

The name of the table resource. A table is a metadata definition that represents your data. You can Grant and Revoke table privileges to a principal.

TaggedDatabase

taggedDatabase_database :: Lens' TaggedDatabase (Maybe DatabaseResource)

A database that has tags attached to it.

taggedDatabase_lFTags :: Lens' TaggedDatabase (Maybe (NonEmpty LFTagPair))

A list of tags attached to the database.

TaggedTable

taggedTable_lFTagsOnTable :: Lens' TaggedTable (Maybe (NonEmpty LFTagPair))

A list of tags attached to the table.

taggedTable_lFTagOnDatabase :: Lens' TaggedTable (Maybe (NonEmpty LFTagPair))

A list of tags attached to the database where the table resides.

taggedTable_lFTagsOnColumns :: Lens' TaggedTable (Maybe [ColumnLFTag])

A list of tags attached to columns in the table.

taggedTable_table :: Lens' TaggedTable (Maybe TableResource)

A table that has tags attached to it.