Amazonka.LakeFormation.ListResources

Description

Lists the resources registered to be managed by the Data Catalog.

Creating a Request

data ListResources

See: newListResources smart constructor.

Constructors

ListResources'

Fields

filterConditionList :: Maybe (NonEmpty FilterCondition)

Any applicable row-level and/or column-level filtering conditions for the resources.

nextToken :: Maybe Text

A continuation token, if this is not the first call to retrieve these resources.

maxResults :: Maybe Natural

The maximum number of resource results.

newListResources :: ListResources

Create a value of ListResources with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:filterConditionList:ListResources', listResources_filterConditionList - Any applicable row-level and/or column-level filtering conditions for the resources.

$sel:nextToken:ListResources', listResources_nextToken - A continuation token, if this is not the first call to retrieve these resources.

$sel:maxResults:ListResources', listResources_maxResults - The maximum number of resource results.

Request Lenses

listResources_filterConditionList :: Lens' ListResources (Maybe (NonEmpty FilterCondition))

Any applicable row-level and/or column-level filtering conditions for the resources.

listResources_nextToken :: Lens' ListResources (Maybe Text)

A continuation token, if this is not the first call to retrieve these resources.

listResources_maxResults :: Lens' ListResources (Maybe Natural)

The maximum number of resource results.

Destructuring the Response

data ListResourcesResponse

See: newListResourcesResponse smart constructor.

Constructors

ListResourcesResponse'

Fields

resourceInfoList :: Maybe [ResourceInfo]

A summary of the data lake resources.

nextToken :: Maybe Text

A continuation token, if this is not the first call to retrieve these resources.

httpStatus :: Int

The response's http status code.

newListResourcesResponse

Arguments

:: Int
$sel:httpStatus:ListResourcesResponse'

-> ListResourcesResponse

Create a value of ListResourcesResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:resourceInfoList:ListResourcesResponse', listResourcesResponse_resourceInfoList - A summary of the data lake resources.

$sel:nextToken:ListResources', listResourcesResponse_nextToken - A continuation token, if this is not the first call to retrieve these resources.

$sel:httpStatus:ListResourcesResponse', listResourcesResponse_httpStatus - The response's http status code.

Response Lenses

listResourcesResponse_resourceInfoList :: Lens' ListResourcesResponse (Maybe [ResourceInfo])

A summary of the data lake resources.

listResourcesResponse_nextToken :: Lens' ListResourcesResponse (Maybe Text)

A continuation token, if this is not the first call to retrieve these resources.

listResourcesResponse_httpStatus :: Lens' ListResourcesResponse Int

The response's http status code.