Amazonka.LakeFormation.RemoveLFTagsFromResource

Description

Removes a tag from the resource. Only database, table, or tableWithColumns resource are allowed. To tag columns, use the column inclusion list in tableWithColumns to specify column input.

Creating a Request

data RemoveLFTagsFromResource

See: newRemoveLFTagsFromResource smart constructor.

Constructors

RemoveLFTagsFromResource'

Fields

catalogId :: Maybe Text

The identifier for the Data Catalog. By default, the account ID. The Data Catalog is the persistent metadata store. It contains database definitions, table definitions, and other control information to manage your AWS Lake Formation environment.

resource :: Resource

The resource where you want to remove a tag.

lFTags :: NonEmpty LFTagPair

The tags to be removed from the resource.

newRemoveLFTagsFromResource

Arguments

:: Resource
$sel:resource:RemoveLFTagsFromResource'

-> NonEmpty LFTagPair
$sel:lFTags:RemoveLFTagsFromResource'

-> RemoveLFTagsFromResource

Create a value of RemoveLFTagsFromResource with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:catalogId:RemoveLFTagsFromResource', removeLFTagsFromResource_catalogId - The identifier for the Data Catalog. By default, the account ID. The Data Catalog is the persistent metadata store. It contains database definitions, table definitions, and other control information to manage your AWS Lake Formation environment.

$sel:resource:RemoveLFTagsFromResource', removeLFTagsFromResource_resource - The resource where you want to remove a tag.

$sel:lFTags:RemoveLFTagsFromResource', removeLFTagsFromResource_lFTags - The tags to be removed from the resource.

Request Lenses

removeLFTagsFromResource_catalogId :: Lens' RemoveLFTagsFromResource (Maybe Text)

The identifier for the Data Catalog. By default, the account ID. The Data Catalog is the persistent metadata store. It contains database definitions, table definitions, and other control information to manage your AWS Lake Formation environment.

removeLFTagsFromResource_resource :: Lens' RemoveLFTagsFromResource Resource

The resource where you want to remove a tag.

removeLFTagsFromResource_lFTags :: Lens' RemoveLFTagsFromResource (NonEmpty LFTagPair)

The tags to be removed from the resource.

Destructuring the Response

data RemoveLFTagsFromResourceResponse

See: newRemoveLFTagsFromResourceResponse smart constructor.

Constructors

RemoveLFTagsFromResourceResponse'

Fields

failures :: Maybe [LFTagError]

A list of failures to untag a resource.

httpStatus :: Int

The response's http status code.

newRemoveLFTagsFromResourceResponse

Arguments

:: Int
$sel:httpStatus:RemoveLFTagsFromResourceResponse'

-> RemoveLFTagsFromResourceResponse

Create a value of RemoveLFTagsFromResourceResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:failures:RemoveLFTagsFromResourceResponse', removeLFTagsFromResourceResponse_failures - A list of failures to untag a resource.

$sel:httpStatus:RemoveLFTagsFromResourceResponse', removeLFTagsFromResourceResponse_httpStatus - The response's http status code.

Response Lenses

removeLFTagsFromResourceResponse_failures :: Lens' RemoveLFTagsFromResourceResponse (Maybe [LFTagError])

A list of failures to untag a resource.

removeLFTagsFromResourceResponse_httpStatus :: Lens' RemoveLFTagsFromResourceResponse Int

The response's http status code.