Amazonka.LakeFormation.RevokePermissions

Description

Revokes permissions to the principal to access metadata in the Data Catalog and data organized in underlying data storage such as Amazon S3.

Creating a Request

data RevokePermissions

See: newRevokePermissions smart constructor.

Constructors

RevokePermissions'

Fields

catalogId :: Maybe Text

The identifier for the Data Catalog. By default, the account ID. The Data Catalog is the persistent metadata store. It contains database definitions, table definitions, and other control information to manage your AWS Lake Formation environment.

permissionsWithGrantOption :: Maybe [Permission]

Indicates a list of permissions for which to revoke the grant option allowing the principal to pass permissions to other principals.

principal :: DataLakePrincipal

The principal to be revoked permissions on the resource.

resource :: Resource

The resource to which permissions are to be revoked.

permissions :: [Permission]

The permissions revoked to the principal on the resource. For information about permissions, see Security and Access Control to Metadata and Data.

newRevokePermissions

Arguments

:: DataLakePrincipal
$sel:principal:RevokePermissions'

-> Resource
$sel:resource:RevokePermissions'

-> RevokePermissions

Create a value of RevokePermissions with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:catalogId:RevokePermissions', revokePermissions_catalogId - The identifier for the Data Catalog. By default, the account ID. The Data Catalog is the persistent metadata store. It contains database definitions, table definitions, and other control information to manage your AWS Lake Formation environment.

$sel:permissionsWithGrantOption:RevokePermissions', revokePermissions_permissionsWithGrantOption - Indicates a list of permissions for which to revoke the grant option allowing the principal to pass permissions to other principals.

$sel:principal:RevokePermissions', revokePermissions_principal - The principal to be revoked permissions on the resource.

$sel:resource:RevokePermissions', revokePermissions_resource - The resource to which permissions are to be revoked.

$sel:permissions:RevokePermissions', revokePermissions_permissions - The permissions revoked to the principal on the resource. For information about permissions, see Security and Access Control to Metadata and Data.

Request Lenses

revokePermissions_catalogId :: Lens' RevokePermissions (Maybe Text)

The identifier for the Data Catalog. By default, the account ID. The Data Catalog is the persistent metadata store. It contains database definitions, table definitions, and other control information to manage your AWS Lake Formation environment.

revokePermissions_permissionsWithGrantOption :: Lens' RevokePermissions (Maybe [Permission])

Indicates a list of permissions for which to revoke the grant option allowing the principal to pass permissions to other principals.

revokePermissions_principal :: Lens' RevokePermissions DataLakePrincipal

The principal to be revoked permissions on the resource.

revokePermissions_resource :: Lens' RevokePermissions Resource

The resource to which permissions are to be revoked.

revokePermissions_permissions :: Lens' RevokePermissions [Permission]

The permissions revoked to the principal on the resource. For information about permissions, see Security and Access Control to Metadata and Data.

Destructuring the Response

data RevokePermissionsResponse

See: newRevokePermissionsResponse smart constructor.

Constructors

RevokePermissionsResponse'

Fields

httpStatus :: Int

The response's http status code.

newRevokePermissionsResponse

Arguments

:: Int
$sel:httpStatus:RevokePermissionsResponse'

-> RevokePermissionsResponse

Create a value of RevokePermissionsResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:httpStatus:RevokePermissionsResponse', revokePermissionsResponse_httpStatus - The response's http status code.

Response Lenses

revokePermissionsResponse_httpStatus :: Lens' RevokePermissionsResponse Int

The response's http status code.