Amazonka.LakeFormation.Types.FilterCondition

Description

Documentation

data FilterCondition

This structure describes the filtering of columns in a table based on a filter condition.

See: newFilterCondition smart constructor.

Constructors

FilterCondition'

Fields

field :: Maybe FieldNameString

The field to filter in the filter condition.

comparisonOperator :: Maybe ComparisonOperator

The comparison operator used in the filter condition.

stringValueList :: Maybe [Text]

A string with values used in evaluating the filter condition.

newFilterCondition :: FilterCondition

Create a value of FilterCondition with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:field:FilterCondition', filterCondition_field - The field to filter in the filter condition.

$sel:comparisonOperator:FilterCondition', filterCondition_comparisonOperator - The comparison operator used in the filter condition.

$sel:stringValueList:FilterCondition', filterCondition_stringValueList - A string with values used in evaluating the filter condition.

filterCondition_field :: Lens' FilterCondition (Maybe FieldNameString)

The field to filter in the filter condition.

filterCondition_comparisonOperator :: Lens' FilterCondition (Maybe ComparisonOperator)

The comparison operator used in the filter condition.

filterCondition_stringValueList :: Lens' FilterCondition (Maybe [Text])

A string with values used in evaluating the filter condition.