Amazonka.Lambda.GetEventSourceMapping

Description

Returns details about an event source mapping. You can get the identifier of a mapping from the output of ListEventSourceMappings.

Creating a Request

data GetEventSourceMapping

See: newGetEventSourceMapping smart constructor.

Constructors

GetEventSourceMapping'

Fields

uuid :: Text

The identifier of the event source mapping.

newGetEventSourceMapping

Arguments

:: Text
$sel:uuid:GetEventSourceMapping'

-> GetEventSourceMapping

Create a value of GetEventSourceMapping with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:uuid:GetEventSourceMapping', getEventSourceMapping_uuid - The identifier of the event source mapping.

Request Lenses

getEventSourceMapping_uuid :: Lens' GetEventSourceMapping Text

The identifier of the event source mapping.

Destructuring the Response

data EventSourceMappingConfiguration

A mapping between an Amazon Web Services resource and a Lambda function. For details, see CreateEventSourceMapping.

See: newEventSourceMappingConfiguration smart constructor.

Constructors

EventSourceMappingConfiguration'

Fields

eventSourceArn :: Maybe Text

The Amazon Resource Name (ARN) of the event source.

state :: Maybe Text

The state of the event source mapping. It can be one of the following: Creating, Enabling, Enabled, Disabling, Disabled, Updating, or Deleting.

startingPositionTimestamp :: Maybe POSIX

With StartingPosition set to AT_TIMESTAMP, the time from which to start reading.

functionArn :: Maybe Text

The ARN of the Lambda function.

topics :: Maybe (NonEmpty Text)

The name of the Kafka topic.

queues :: Maybe (NonEmpty Text)

(Amazon MQ) The name of the Amazon MQ broker destination queue to consume.

bisectBatchOnFunctionError :: Maybe Bool

(Streams only) If the function returns an error, split the batch in two and retry. The default value is false.

uuid :: Maybe Text

The identifier of the event source mapping.

parallelizationFactor :: Maybe Natural

(Streams only) The number of batches to process concurrently from each shard. The default value is 1.

lastProcessingResult :: Maybe Text

The result of the last Lambda invocation of your function.

maximumRetryAttempts :: Maybe Int

(Streams only) Discard records after the specified number of retries. The default value is -1, which sets the maximum number of retries to infinite. When MaximumRetryAttempts is infinite, Lambda retries failed records until the record expires in the event source.

batchSize :: Maybe Natural

The maximum number of records in each batch that Lambda pulls from your stream or queue and sends to your function. Lambda passes all of the records in the batch to the function in a single call, up to the payload limit for synchronous invocation (6 MB).

Default value: Varies by service. For Amazon SQS, the default is 10. For all other services, the default is 100.

Related setting: When you set BatchSize to a value greater than 10, you must set MaximumBatchingWindowInSeconds to at least 1.

stateTransitionReason :: Maybe Text

Indicates whether a user or Lambda made the last change to the event source mapping.

maximumBatchingWindowInSeconds :: Maybe Natural

(Streams and Amazon SQS standard queues) The maximum amount of time, in seconds, that Lambda spends gathering records before invoking the function.

Default: 0

Related setting: When you set BatchSize to a value greater than 10, you must set MaximumBatchingWindowInSeconds to at least 1.

sourceAccessConfigurations :: Maybe [SourceAccessConfiguration]

An array of the authentication protocol, VPC components, or virtual host to secure and define your event source.

maximumRecordAgeInSeconds :: Maybe Int

(Streams only) Discard records older than the specified age. The default value is -1, which sets the maximum age to infinite. When the value is set to infinite, Lambda never discards old records.

functionResponseTypes :: Maybe [FunctionResponseType]

(Streams only) A list of current response type enums applied to the event source mapping.

tumblingWindowInSeconds :: Maybe Natural

(Streams only) The duration in seconds of a processing window. The range is 1–900 seconds.

selfManagedEventSource :: Maybe SelfManagedEventSource

The self-managed Apache Kafka cluster for your event source.

lastModified :: Maybe POSIX

The date that the event source mapping was last updated or that its state changed.

destinationConfig :: Maybe DestinationConfig

(Streams only) An Amazon SQS queue or Amazon SNS topic destination for discarded records.

startingPosition :: Maybe EventSourcePosition

The position in a stream from which to start reading. Required for Amazon Kinesis, Amazon DynamoDB, and Amazon MSK stream sources. AT_TIMESTAMP is supported only for Amazon Kinesis streams.

newEventSourceMappingConfiguration :: EventSourceMappingConfiguration

Create a value of EventSourceMappingConfiguration with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:eventSourceArn:EventSourceMappingConfiguration', eventSourceMappingConfiguration_eventSourceArn - The Amazon Resource Name (ARN) of the event source.

$sel:state:EventSourceMappingConfiguration', eventSourceMappingConfiguration_state - The state of the event source mapping. It can be one of the following: Creating, Enabling, Enabled, Disabling, Disabled, Updating, or Deleting.

$sel:startingPositionTimestamp:EventSourceMappingConfiguration', eventSourceMappingConfiguration_startingPositionTimestamp - With StartingPosition set to AT_TIMESTAMP, the time from which to start reading.

$sel:functionArn:EventSourceMappingConfiguration', eventSourceMappingConfiguration_functionArn - The ARN of the Lambda function.

$sel:topics:EventSourceMappingConfiguration', eventSourceMappingConfiguration_topics - The name of the Kafka topic.

$sel:queues:EventSourceMappingConfiguration', eventSourceMappingConfiguration_queues - (Amazon MQ) The name of the Amazon MQ broker destination queue to consume.

$sel:bisectBatchOnFunctionError:EventSourceMappingConfiguration', eventSourceMappingConfiguration_bisectBatchOnFunctionError - (Streams only) If the function returns an error, split the batch in two and retry. The default value is false.

$sel:uuid:EventSourceMappingConfiguration', eventSourceMappingConfiguration_uuid - The identifier of the event source mapping.

$sel:parallelizationFactor:EventSourceMappingConfiguration', eventSourceMappingConfiguration_parallelizationFactor - (Streams only) The number of batches to process concurrently from each shard. The default value is 1.

$sel:lastProcessingResult:EventSourceMappingConfiguration', eventSourceMappingConfiguration_lastProcessingResult - The result of the last Lambda invocation of your function.

$sel:maximumRetryAttempts:EventSourceMappingConfiguration', eventSourceMappingConfiguration_maximumRetryAttempts - (Streams only) Discard records after the specified number of retries. The default value is -1, which sets the maximum number of retries to infinite. When MaximumRetryAttempts is infinite, Lambda retries failed records until the record expires in the event source.

$sel:batchSize:EventSourceMappingConfiguration', eventSourceMappingConfiguration_batchSize - The maximum number of records in each batch that Lambda pulls from your stream or queue and sends to your function. Lambda passes all of the records in the batch to the function in a single call, up to the payload limit for synchronous invocation (6 MB).

Default value: Varies by service. For Amazon SQS, the default is 10. For all other services, the default is 100.

Related setting: When you set BatchSize to a value greater than 10, you must set MaximumBatchingWindowInSeconds to at least 1.

$sel:stateTransitionReason:EventSourceMappingConfiguration', eventSourceMappingConfiguration_stateTransitionReason - Indicates whether a user or Lambda made the last change to the event source mapping.

$sel:maximumBatchingWindowInSeconds:EventSourceMappingConfiguration', eventSourceMappingConfiguration_maximumBatchingWindowInSeconds - (Streams and Amazon SQS standard queues) The maximum amount of time, in seconds, that Lambda spends gathering records before invoking the function.

Default: 0

Related setting: When you set BatchSize to a value greater than 10, you must set MaximumBatchingWindowInSeconds to at least 1.

$sel:sourceAccessConfigurations:EventSourceMappingConfiguration', eventSourceMappingConfiguration_sourceAccessConfigurations - An array of the authentication protocol, VPC components, or virtual host to secure and define your event source.

$sel:maximumRecordAgeInSeconds:EventSourceMappingConfiguration', eventSourceMappingConfiguration_maximumRecordAgeInSeconds - (Streams only) Discard records older than the specified age. The default value is -1, which sets the maximum age to infinite. When the value is set to infinite, Lambda never discards old records.

$sel:functionResponseTypes:EventSourceMappingConfiguration', eventSourceMappingConfiguration_functionResponseTypes - (Streams only) A list of current response type enums applied to the event source mapping.

$sel:tumblingWindowInSeconds:EventSourceMappingConfiguration', eventSourceMappingConfiguration_tumblingWindowInSeconds - (Streams only) The duration in seconds of a processing window. The range is 1–900 seconds.

$sel:selfManagedEventSource:EventSourceMappingConfiguration', eventSourceMappingConfiguration_selfManagedEventSource - The self-managed Apache Kafka cluster for your event source.

$sel:lastModified:EventSourceMappingConfiguration', eventSourceMappingConfiguration_lastModified - The date that the event source mapping was last updated or that its state changed.

$sel:destinationConfig:EventSourceMappingConfiguration', eventSourceMappingConfiguration_destinationConfig - (Streams only) An Amazon SQS queue or Amazon SNS topic destination for discarded records.

$sel:startingPosition:EventSourceMappingConfiguration', eventSourceMappingConfiguration_startingPosition - The position in a stream from which to start reading. Required for Amazon Kinesis, Amazon DynamoDB, and Amazon MSK stream sources. AT_TIMESTAMP is supported only for Amazon Kinesis streams.

Response Lenses

eventSourceMappingConfiguration_eventSourceArn :: Lens' EventSourceMappingConfiguration (Maybe Text)

The Amazon Resource Name (ARN) of the event source.

eventSourceMappingConfiguration_state :: Lens' EventSourceMappingConfiguration (Maybe Text)

The state of the event source mapping. It can be one of the following: Creating, Enabling, Enabled, Disabling, Disabled, Updating, or Deleting.

eventSourceMappingConfiguration_startingPositionTimestamp :: Lens' EventSourceMappingConfiguration (Maybe UTCTime)

With StartingPosition set to AT_TIMESTAMP, the time from which to start reading.

eventSourceMappingConfiguration_functionArn :: Lens' EventSourceMappingConfiguration (Maybe Text)

The ARN of the Lambda function.

eventSourceMappingConfiguration_topics :: Lens' EventSourceMappingConfiguration (Maybe (NonEmpty Text))

The name of the Kafka topic.

eventSourceMappingConfiguration_queues :: Lens' EventSourceMappingConfiguration (Maybe (NonEmpty Text))

(Amazon MQ) The name of the Amazon MQ broker destination queue to consume.

eventSourceMappingConfiguration_bisectBatchOnFunctionError :: Lens' EventSourceMappingConfiguration (Maybe Bool)

(Streams only) If the function returns an error, split the batch in two and retry. The default value is false.

eventSourceMappingConfiguration_uuid :: Lens' EventSourceMappingConfiguration (Maybe Text)

The identifier of the event source mapping.

eventSourceMappingConfiguration_parallelizationFactor :: Lens' EventSourceMappingConfiguration (Maybe Natural)

(Streams only) The number of batches to process concurrently from each shard. The default value is 1.

eventSourceMappingConfiguration_lastProcessingResult :: Lens' EventSourceMappingConfiguration (Maybe Text)

The result of the last Lambda invocation of your function.

eventSourceMappingConfiguration_maximumRetryAttempts :: Lens' EventSourceMappingConfiguration (Maybe Int)

(Streams only) Discard records after the specified number of retries. The default value is -1, which sets the maximum number of retries to infinite. When MaximumRetryAttempts is infinite, Lambda retries failed records until the record expires in the event source.

eventSourceMappingConfiguration_batchSize :: Lens' EventSourceMappingConfiguration (Maybe Natural)

The maximum number of records in each batch that Lambda pulls from your stream or queue and sends to your function. Lambda passes all of the records in the batch to the function in a single call, up to the payload limit for synchronous invocation (6 MB).

Default value: Varies by service. For Amazon SQS, the default is 10. For all other services, the default is 100.

Related setting: When you set BatchSize to a value greater than 10, you must set MaximumBatchingWindowInSeconds to at least 1.

eventSourceMappingConfiguration_stateTransitionReason :: Lens' EventSourceMappingConfiguration (Maybe Text)

Indicates whether a user or Lambda made the last change to the event source mapping.

eventSourceMappingConfiguration_maximumBatchingWindowInSeconds :: Lens' EventSourceMappingConfiguration (Maybe Natural)

(Streams and Amazon SQS standard queues) The maximum amount of time, in seconds, that Lambda spends gathering records before invoking the function.

Default: 0

Related setting: When you set BatchSize to a value greater than 10, you must set MaximumBatchingWindowInSeconds to at least 1.

eventSourceMappingConfiguration_sourceAccessConfigurations :: Lens' EventSourceMappingConfiguration (Maybe [SourceAccessConfiguration])

An array of the authentication protocol, VPC components, or virtual host to secure and define your event source.

eventSourceMappingConfiguration_maximumRecordAgeInSeconds :: Lens' EventSourceMappingConfiguration (Maybe Int)

(Streams only) Discard records older than the specified age. The default value is -1, which sets the maximum age to infinite. When the value is set to infinite, Lambda never discards old records.

eventSourceMappingConfiguration_functionResponseTypes :: Lens' EventSourceMappingConfiguration (Maybe [FunctionResponseType])

(Streams only) A list of current response type enums applied to the event source mapping.

eventSourceMappingConfiguration_tumblingWindowInSeconds :: Lens' EventSourceMappingConfiguration (Maybe Natural)

(Streams only) The duration in seconds of a processing window. The range is 1–900 seconds.

eventSourceMappingConfiguration_selfManagedEventSource :: Lens' EventSourceMappingConfiguration (Maybe SelfManagedEventSource)

The self-managed Apache Kafka cluster for your event source.

eventSourceMappingConfiguration_lastModified :: Lens' EventSourceMappingConfiguration (Maybe UTCTime)

The date that the event source mapping was last updated or that its state changed.

eventSourceMappingConfiguration_destinationConfig :: Lens' EventSourceMappingConfiguration (Maybe DestinationConfig)

(Streams only) An Amazon SQS queue or Amazon SNS topic destination for discarded records.

eventSourceMappingConfiguration_startingPosition :: Lens' EventSourceMappingConfiguration (Maybe EventSourcePosition)

The position in a stream from which to start reading. Required for Amazon Kinesis, Amazon DynamoDB, and Amazon MSK stream sources. AT_TIMESTAMP is supported only for Amazon Kinesis streams.