Amazonka.Lambda.GetLayerVersionByArn

Description

Returns information about a version of an Lambda layer, with a link to download the layer archive that's valid for 10 minutes.

Creating a Request

data GetLayerVersionByArn

See: newGetLayerVersionByArn smart constructor.

Constructors

GetLayerVersionByArn'

Fields

arn :: Text

The ARN of the layer version.

newGetLayerVersionByArn

Arguments

:: Text
$sel:arn:GetLayerVersionByArn'

-> GetLayerVersionByArn

Create a value of GetLayerVersionByArn with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:arn:GetLayerVersionByArn', getLayerVersionByArn_arn - The ARN of the layer version.

Request Lenses

getLayerVersionByArn_arn :: Lens' GetLayerVersionByArn Text

The ARN of the layer version.

Destructuring the Response

data GetLayerVersionResponse

See: newGetLayerVersionResponse smart constructor.

Constructors

GetLayerVersionResponse'

Fields

layerVersionArn :: Maybe Text

The ARN of the layer version.

content :: Maybe LayerVersionContentOutput

Details about the layer version.

createdDate :: Maybe Text

The date that the layer version was created, in ISO-8601 format (YYYY-MM-DDThh:mm:ss.sTZD).

version :: Maybe Integer

The version number.

licenseInfo :: Maybe Text

The layer's software license.

compatibleArchitectures :: Maybe [Architecture]

A list of compatible instruction set architectures.

layerArn :: Maybe Text

The ARN of the layer.

description :: Maybe Text

The description of the version.

compatibleRuntimes :: Maybe [Runtime]

The layer's compatible runtimes.

newGetLayerVersionResponse :: GetLayerVersionResponse

Create a value of GetLayerVersionResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:layerVersionArn:GetLayerVersionResponse', getLayerVersionResponse_layerVersionArn - The ARN of the layer version.

$sel:content:GetLayerVersionResponse', getLayerVersionResponse_content - Details about the layer version.

$sel:createdDate:GetLayerVersionResponse', getLayerVersionResponse_createdDate - The date that the layer version was created, in ISO-8601 format (YYYY-MM-DDThh:mm:ss.sTZD).

$sel:version:GetLayerVersionResponse', getLayerVersionResponse_version - The version number.

$sel:licenseInfo:GetLayerVersionResponse', getLayerVersionResponse_licenseInfo - The layer's software license.

$sel:compatibleArchitectures:GetLayerVersionResponse', getLayerVersionResponse_compatibleArchitectures - A list of compatible instruction set architectures.

$sel:layerArn:GetLayerVersionResponse', getLayerVersionResponse_layerArn - The ARN of the layer.

$sel:description:GetLayerVersionResponse', getLayerVersionResponse_description - The description of the version.

$sel:compatibleRuntimes:GetLayerVersionResponse', getLayerVersionResponse_compatibleRuntimes - The layer's compatible runtimes.

Response Lenses

getLayerVersionResponse_layerVersionArn :: Lens' GetLayerVersionResponse (Maybe Text)

The ARN of the layer version.

getLayerVersionResponse_content :: Lens' GetLayerVersionResponse (Maybe LayerVersionContentOutput)

Details about the layer version.

getLayerVersionResponse_createdDate :: Lens' GetLayerVersionResponse (Maybe Text)

The date that the layer version was created, in ISO-8601 format (YYYY-MM-DDThh:mm:ss.sTZD).

getLayerVersionResponse_version :: Lens' GetLayerVersionResponse (Maybe Integer)

The version number.

getLayerVersionResponse_licenseInfo :: Lens' GetLayerVersionResponse (Maybe Text)

The layer's software license.

getLayerVersionResponse_compatibleArchitectures :: Lens' GetLayerVersionResponse (Maybe [Architecture])

A list of compatible instruction set architectures.

getLayerVersionResponse_layerArn :: Lens' GetLayerVersionResponse (Maybe Text)

The ARN of the layer.

getLayerVersionResponse_description :: Lens' GetLayerVersionResponse (Maybe Text)

The description of the version.

getLayerVersionResponse_compatibleRuntimes :: Lens' GetLayerVersionResponse (Maybe [Runtime])

The layer's compatible runtimes.