Amazonka.Lambda.Invoke

Description

Invokes a Lambda function. You can invoke a function synchronously (and wait for the response), or asynchronously. To invoke a function asynchronously, set InvocationType to Event.

For synchronous invocation, details about the function response, including errors, are included in the response body and headers. For either invocation type, you can find more information in the execution log and trace.

When an error occurs, your function may be invoked multiple times. Retry behavior varies by error type, client, event source, and invocation type. For example, if you invoke a function asynchronously and it returns an error, Lambda executes the function up to two more times. For more information, see Retry Behavior.

For asynchronous invocation, Lambda adds events to a queue before sending them to your function. If your function does not have enough capacity to keep up with the queue, events may be lost. Occasionally, your function may receive the same event multiple times, even if no error occurs. To retain events that were not processed, configure your function with a dead-letter queue.

The status code in the API response doesn't reflect function errors. Error codes are reserved for errors that prevent your function from executing, such as permissions errors, limit errors, or issues with your function's code and configuration. For example, Lambda returns TooManyRequestsException if executing the function would cause you to exceed a concurrency limit at either the account level (ConcurrentInvocationLimitExceeded) or function level (ReservedFunctionConcurrentInvocationLimitExceeded).

For functions with a long timeout, your client might be disconnected during synchronous invocation while it waits for a response. Configure your HTTP client, SDK, firewall, proxy, or operating system to allow for long connections with timeout or keep-alive settings.

This operation requires permission for the lambda:InvokeFunction action.

Creating a Request

data Invoke

See: newInvoke smart constructor.

Constructors

Invoke'

Fields

invocationType :: Maybe InvocationType

Choose from the following options.

RequestResponse (default) - Invoke the function synchronously. Keep the connection open until the function returns a response or times out. The API response includes the function response and additional data.

Event - Invoke the function asynchronously. Send events that fail multiple times to the function's dead-letter queue (if it's configured). The API response only includes a status code.

DryRun - Validate parameter values and verify that the user or role has permission to invoke the function.

logType :: Maybe LogType

Set to Tail to include the execution log in the response. Applies to synchronously invoked functions only.

qualifier :: Maybe Text

Specify a version or alias to invoke a published version of the function.

clientContext :: Maybe Text

Up to 3583 bytes of base64-encoded data about the invoking client to pass to the function in the context object.

functionName :: Text

The name of the Lambda function, version, or alias.

Name formats

Function name - my-function (name-only), my-function:v1 (with alias).

Function ARN - arn:aws:lambda:us-west-2:123456789012:function:my-function.

Partial ARN - 123456789012:function:my-function.

You can append a version number or alias to any of the formats. The length constraint applies only to the full ARN. If you specify only the function name, it is limited to 64 characters in length.

payload :: ByteString

The JSON that you want to provide to your Lambda function as input.

newInvoke

Arguments

:: Text
$sel:functionName:Invoke'

-> ByteString
$sel:payload:Invoke'

-> Invoke

Create a value of Invoke with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:invocationType:Invoke', invoke_invocationType - Choose from the following options.

RequestResponse (default) - Invoke the function synchronously. Keep the connection open until the function returns a response or times out. The API response includes the function response and additional data.

Event - Invoke the function asynchronously. Send events that fail multiple times to the function's dead-letter queue (if it's configured). The API response only includes a status code.

DryRun - Validate parameter values and verify that the user or role has permission to invoke the function.

$sel:logType:Invoke', invoke_logType - Set to Tail to include the execution log in the response. Applies to synchronously invoked functions only.

$sel:qualifier:Invoke', invoke_qualifier - Specify a version or alias to invoke a published version of the function.

$sel:clientContext:Invoke', invoke_clientContext - Up to 3583 bytes of base64-encoded data about the invoking client to pass to the function in the context object.

$sel:functionName:Invoke', invoke_functionName - The name of the Lambda function, version, or alias.

Name formats

Function name - my-function (name-only), my-function:v1 (with alias).

Function ARN - arn:aws:lambda:us-west-2:123456789012:function:my-function.

Partial ARN - 123456789012:function:my-function.

You can append a version number or alias to any of the formats. The length constraint applies only to the full ARN. If you specify only the function name, it is limited to 64 characters in length.

$sel:payload:Invoke', invoke_payload - The JSON that you want to provide to your Lambda function as input.

Request Lenses

invoke_invocationType :: Lens' Invoke (Maybe InvocationType)

Choose from the following options.

RequestResponse (default) - Invoke the function synchronously. Keep the connection open until the function returns a response or times out. The API response includes the function response and additional data.

Event - Invoke the function asynchronously. Send events that fail multiple times to the function's dead-letter queue (if it's configured). The API response only includes a status code.

DryRun - Validate parameter values and verify that the user or role has permission to invoke the function.

invoke_logType :: Lens' Invoke (Maybe LogType)

Set to Tail to include the execution log in the response. Applies to synchronously invoked functions only.

invoke_qualifier :: Lens' Invoke (Maybe Text)

Specify a version or alias to invoke a published version of the function.

invoke_clientContext :: Lens' Invoke (Maybe Text)

Up to 3583 bytes of base64-encoded data about the invoking client to pass to the function in the context object.

invoke_functionName :: Lens' Invoke Text

The name of the Lambda function, version, or alias.

Name formats

Function name - my-function (name-only), my-function:v1 (with alias).

Function ARN - arn:aws:lambda:us-west-2:123456789012:function:my-function.

Partial ARN - 123456789012:function:my-function.

You can append a version number or alias to any of the formats. The length constraint applies only to the full ARN. If you specify only the function name, it is limited to 64 characters in length.

invoke_payload :: Lens' Invoke ByteString

The JSON that you want to provide to your Lambda function as input.

Destructuring the Response

data InvokeResponse

See: newInvokeResponse smart constructor.

Constructors

InvokeResponse'

Fields

functionError :: Maybe Text

If present, indicates that an error occurred during function execution. Details about the error are included in the response payload.

logResult :: Maybe Text

The last 4 KB of the execution log, which is base64 encoded.

payload :: Maybe ByteString

The response from the function, or an error object.

executedVersion :: Maybe Text

The version of the function that executed. When you invoke a function with an alias, this indicates which version the alias resolved to.

statusCode :: Int

The HTTP status code is in the 200 range for a successful request. For the RequestResponse invocation type, this status code is 200. For the Event invocation type, this status code is 202. For the DryRun invocation type, the status code is 204.

newInvokeResponse

Arguments

:: Int
$sel:statusCode:InvokeResponse'

-> InvokeResponse

Create a value of InvokeResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:functionError:InvokeResponse', invokeResponse_functionError - If present, indicates that an error occurred during function execution. Details about the error are included in the response payload.

$sel:logResult:InvokeResponse', invokeResponse_logResult - The last 4 KB of the execution log, which is base64 encoded.

$sel:payload:Invoke', invokeResponse_payload - The response from the function, or an error object.

$sel:executedVersion:InvokeResponse', invokeResponse_executedVersion - The version of the function that executed. When you invoke a function with an alias, this indicates which version the alias resolved to.

$sel:statusCode:InvokeResponse', invokeResponse_statusCode - The HTTP status code is in the 200 range for a successful request. For the RequestResponse invocation type, this status code is 200. For the Event invocation type, this status code is 202. For the DryRun invocation type, the status code is 204.

Response Lenses

invokeResponse_functionError :: Lens' InvokeResponse (Maybe Text)

If present, indicates that an error occurred during function execution. Details about the error are included in the response payload.

invokeResponse_logResult :: Lens' InvokeResponse (Maybe Text)

The last 4 KB of the execution log, which is base64 encoded.

invokeResponse_payload :: Lens' InvokeResponse (Maybe ByteString)

The response from the function, or an error object.

invokeResponse_executedVersion :: Lens' InvokeResponse (Maybe Text)

The version of the function that executed. When you invoke a function with an alias, this indicates which version the alias resolved to.

invokeResponse_statusCode :: Lens' InvokeResponse Int

The HTTP status code is in the 200 range for a successful request. For the RequestResponse invocation type, this status code is 200. For the Event invocation type, this status code is 202. For the DryRun invocation type, the status code is 204.