Amazonka.Lambda.ListCodeSigningConfigs

Description

Returns a list of code signing configurations. A request returns up to 10,000 configurations per call. You can use the MaxItems parameter to return fewer configurations per call.

This operation returns paginated results.

Creating a Request

data ListCodeSigningConfigs

See: newListCodeSigningConfigs smart constructor.

Constructors

ListCodeSigningConfigs'

Fields

marker :: Maybe Text

Specify the pagination token that's returned by a previous request to retrieve the next page of results.

maxItems :: Maybe Natural

Maximum number of items to return.

newListCodeSigningConfigs :: ListCodeSigningConfigs

Create a value of ListCodeSigningConfigs with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:marker:ListCodeSigningConfigs', listCodeSigningConfigs_marker - Specify the pagination token that's returned by a previous request to retrieve the next page of results.

$sel:maxItems:ListCodeSigningConfigs', listCodeSigningConfigs_maxItems - Maximum number of items to return.

Request Lenses

listCodeSigningConfigs_marker :: Lens' ListCodeSigningConfigs (Maybe Text)

Specify the pagination token that's returned by a previous request to retrieve the next page of results.

listCodeSigningConfigs_maxItems :: Lens' ListCodeSigningConfigs (Maybe Natural)

Maximum number of items to return.

Destructuring the Response

data ListCodeSigningConfigsResponse

See: newListCodeSigningConfigsResponse smart constructor.

Constructors

ListCodeSigningConfigsResponse'

Fields

codeSigningConfigs :: Maybe [CodeSigningConfig]

The code signing configurations

nextMarker :: Maybe Text

The pagination token that's included if more results are available.

httpStatus :: Int

The response's http status code.

newListCodeSigningConfigsResponse

Arguments

:: Int
$sel:httpStatus:ListCodeSigningConfigsResponse'

-> ListCodeSigningConfigsResponse

Create a value of ListCodeSigningConfigsResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:codeSigningConfigs:ListCodeSigningConfigsResponse', listCodeSigningConfigsResponse_codeSigningConfigs - The code signing configurations

$sel:nextMarker:ListCodeSigningConfigsResponse', listCodeSigningConfigsResponse_nextMarker - The pagination token that's included if more results are available.

$sel:httpStatus:ListCodeSigningConfigsResponse', listCodeSigningConfigsResponse_httpStatus - The response's http status code.

Response Lenses

listCodeSigningConfigsResponse_codeSigningConfigs :: Lens' ListCodeSigningConfigsResponse (Maybe [CodeSigningConfig])

The code signing configurations

listCodeSigningConfigsResponse_nextMarker :: Lens' ListCodeSigningConfigsResponse (Maybe Text)

The pagination token that's included if more results are available.

listCodeSigningConfigsResponse_httpStatus :: Lens' ListCodeSigningConfigsResponse Int

The response's http status code.