Amazonka.Lambda.ListVersionsByFunction

Description

Returns a list of versions, with the version-specific configuration of each. Lambda returns up to 50 versions per call.

This operation returns paginated results.

Creating a Request

data ListVersionsByFunction

See: newListVersionsByFunction smart constructor.

Constructors

ListVersionsByFunction'

Fields

marker :: Maybe Text

Specify the pagination token that's returned by a previous request to retrieve the next page of results.

maxItems :: Maybe Natural

The maximum number of versions to return. Note that ListVersionsByFunction returns a maximum of 50 items in each response, even if you set the number higher.

functionName :: Text

The name of the Lambda function.

Name formats

Function name - MyFunction.

Function ARN - arn:aws:lambda:us-west-2:123456789012:function:MyFunction.

Partial ARN - 123456789012:function:MyFunction.

The length constraint applies only to the full ARN. If you specify only the function name, it is limited to 64 characters in length.

newListVersionsByFunction

Arguments

:: Text
$sel:functionName:ListVersionsByFunction'

-> ListVersionsByFunction

Create a value of ListVersionsByFunction with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:marker:ListVersionsByFunction', listVersionsByFunction_marker - Specify the pagination token that's returned by a previous request to retrieve the next page of results.

$sel:maxItems:ListVersionsByFunction', listVersionsByFunction_maxItems - The maximum number of versions to return. Note that ListVersionsByFunction returns a maximum of 50 items in each response, even if you set the number higher.

$sel:functionName:ListVersionsByFunction', listVersionsByFunction_functionName - The name of the Lambda function.

Name formats

Function name - MyFunction.

Function ARN - arn:aws:lambda:us-west-2:123456789012:function:MyFunction.

Partial ARN - 123456789012:function:MyFunction.

The length constraint applies only to the full ARN. If you specify only the function name, it is limited to 64 characters in length.

Request Lenses

listVersionsByFunction_marker :: Lens' ListVersionsByFunction (Maybe Text)

Specify the pagination token that's returned by a previous request to retrieve the next page of results.

listVersionsByFunction_maxItems :: Lens' ListVersionsByFunction (Maybe Natural)

The maximum number of versions to return. Note that ListVersionsByFunction returns a maximum of 50 items in each response, even if you set the number higher.

listVersionsByFunction_functionName :: Lens' ListVersionsByFunction Text

The name of the Lambda function.

Name formats

Function name - MyFunction.

Function ARN - arn:aws:lambda:us-west-2:123456789012:function:MyFunction.

Partial ARN - 123456789012:function:MyFunction.

The length constraint applies only to the full ARN. If you specify only the function name, it is limited to 64 characters in length.

Destructuring the Response

data ListVersionsByFunctionResponse

See: newListVersionsByFunctionResponse smart constructor.

Constructors

ListVersionsByFunctionResponse'

Fields

versions :: Maybe [FunctionConfiguration]

A list of Lambda function versions.

nextMarker :: Maybe Text

The pagination token that's included if more results are available.

httpStatus :: Int

The response's http status code.

newListVersionsByFunctionResponse

Arguments

:: Int
$sel:httpStatus:ListVersionsByFunctionResponse'

-> ListVersionsByFunctionResponse

Create a value of ListVersionsByFunctionResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:versions:ListVersionsByFunctionResponse', listVersionsByFunctionResponse_versions - A list of Lambda function versions.

$sel:nextMarker:ListVersionsByFunctionResponse', listVersionsByFunctionResponse_nextMarker - The pagination token that's included if more results are available.

$sel:httpStatus:ListVersionsByFunctionResponse', listVersionsByFunctionResponse_httpStatus - The response's http status code.

Response Lenses

listVersionsByFunctionResponse_versions :: Lens' ListVersionsByFunctionResponse (Maybe [FunctionConfiguration])

A list of Lambda function versions.

listVersionsByFunctionResponse_nextMarker :: Lens' ListVersionsByFunctionResponse (Maybe Text)

The pagination token that's included if more results are available.

listVersionsByFunctionResponse_httpStatus :: Lens' ListVersionsByFunctionResponse Int

The response's http status code.