Amazonka.Lambda.RemoveLayerVersionPermission

Description

Removes a statement from the permissions policy for a version of an Lambda layer. For more information, see AddLayerVersionPermission.

Creating a Request

data RemoveLayerVersionPermission

See: newRemoveLayerVersionPermission smart constructor.

Constructors

RemoveLayerVersionPermission'

Fields

revisionId :: Maybe Text

Only update the policy if the revision ID matches the ID specified. Use this option to avoid modifying a policy that has changed since you last read it.

layerName :: Text

The name or Amazon Resource Name (ARN) of the layer.

versionNumber :: Integer

The version number.

statementId :: Text

The identifier that was specified when the statement was added.

newRemoveLayerVersionPermission

Arguments

:: Text
$sel:layerName:RemoveLayerVersionPermission'

-> Integer
$sel:versionNumber:RemoveLayerVersionPermission'

-> Text
$sel:statementId:RemoveLayerVersionPermission'

-> RemoveLayerVersionPermission

Create a value of RemoveLayerVersionPermission with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:revisionId:RemoveLayerVersionPermission', removeLayerVersionPermission_revisionId - Only update the policy if the revision ID matches the ID specified. Use this option to avoid modifying a policy that has changed since you last read it.

$sel:layerName:RemoveLayerVersionPermission', removeLayerVersionPermission_layerName - The name or Amazon Resource Name (ARN) of the layer.

$sel:versionNumber:RemoveLayerVersionPermission', removeLayerVersionPermission_versionNumber - The version number.

$sel:statementId:RemoveLayerVersionPermission', removeLayerVersionPermission_statementId - The identifier that was specified when the statement was added.

Request Lenses

removeLayerVersionPermission_revisionId :: Lens' RemoveLayerVersionPermission (Maybe Text)

Only update the policy if the revision ID matches the ID specified. Use this option to avoid modifying a policy that has changed since you last read it.

removeLayerVersionPermission_layerName :: Lens' RemoveLayerVersionPermission Text

The name or Amazon Resource Name (ARN) of the layer.

removeLayerVersionPermission_versionNumber :: Lens' RemoveLayerVersionPermission Integer

The version number.

removeLayerVersionPermission_statementId :: Lens' RemoveLayerVersionPermission Text

The identifier that was specified when the statement was added.

Destructuring the Response

data RemoveLayerVersionPermissionResponse

See: newRemoveLayerVersionPermissionResponse smart constructor.

Constructors

RemoveLayerVersionPermissionResponse'

newRemoveLayerVersionPermissionResponse :: RemoveLayerVersionPermissionResponse

Create a value of RemoveLayerVersionPermissionResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.