Amazonka.Lambda.RemovePermission

Description

Revokes function-use permission from an Amazon Web Services service or another account. You can get the ID of the statement from the output of GetPolicy.

Creating a Request

data RemovePermission

See: newRemovePermission smart constructor.

Constructors

RemovePermission'

Fields

qualifier :: Maybe Text

Specify a version or alias to remove permissions from a published version of the function.

revisionId :: Maybe Text

Only update the policy if the revision ID matches the ID that's specified. Use this option to avoid modifying a policy that has changed since you last read it.

functionName :: Text

The name of the Lambda function, version, or alias.

Name formats

Function name - my-function (name-only), my-function:v1 (with alias).

Function ARN - arn:aws:lambda:us-west-2:123456789012:function:my-function.

Partial ARN - 123456789012:function:my-function.

You can append a version number or alias to any of the formats. The length constraint applies only to the full ARN. If you specify only the function name, it is limited to 64 characters in length.

statementId :: Text

Statement ID of the permission to remove.

newRemovePermission

Arguments

:: Text
$sel:functionName:RemovePermission'

-> Text
$sel:statementId:RemovePermission'

-> RemovePermission

Create a value of RemovePermission with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:qualifier:RemovePermission', removePermission_qualifier - Specify a version or alias to remove permissions from a published version of the function.

$sel:revisionId:RemovePermission', removePermission_revisionId - Only update the policy if the revision ID matches the ID that's specified. Use this option to avoid modifying a policy that has changed since you last read it.

$sel:functionName:RemovePermission', removePermission_functionName - The name of the Lambda function, version, or alias.

Name formats

Function name - my-function (name-only), my-function:v1 (with alias).

Function ARN - arn:aws:lambda:us-west-2:123456789012:function:my-function.

Partial ARN - 123456789012:function:my-function.

You can append a version number or alias to any of the formats. The length constraint applies only to the full ARN. If you specify only the function name, it is limited to 64 characters in length.

$sel:statementId:RemovePermission', removePermission_statementId - Statement ID of the permission to remove.

Request Lenses

removePermission_qualifier :: Lens' RemovePermission (Maybe Text)

Specify a version or alias to remove permissions from a published version of the function.

removePermission_revisionId :: Lens' RemovePermission (Maybe Text)

Only update the policy if the revision ID matches the ID that's specified. Use this option to avoid modifying a policy that has changed since you last read it.

removePermission_functionName :: Lens' RemovePermission Text

The name of the Lambda function, version, or alias.

Name formats

Function name - my-function (name-only), my-function:v1 (with alias).

Function ARN - arn:aws:lambda:us-west-2:123456789012:function:my-function.

Partial ARN - 123456789012:function:my-function.

You can append a version number or alias to any of the formats. The length constraint applies only to the full ARN. If you specify only the function name, it is limited to 64 characters in length.

removePermission_statementId :: Lens' RemovePermission Text

Statement ID of the permission to remove.

Destructuring the Response

data RemovePermissionResponse

See: newRemovePermissionResponse smart constructor.

Constructors

RemovePermissionResponse'

newRemovePermissionResponse :: RemovePermissionResponse

Create a value of RemovePermissionResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.