Amazonka.Lambda.Types.CodeSigningConfig

Description

Documentation

data CodeSigningConfig

Details about a Code signing configuration.

See: newCodeSigningConfig smart constructor.

Constructors

CodeSigningConfig'

Fields

description :: Maybe Text

Code signing configuration description.

codeSigningConfigId :: Text

Unique identifer for the Code signing configuration.

codeSigningConfigArn :: Text

The Amazon Resource Name (ARN) of the Code signing configuration.

allowedPublishers :: AllowedPublishers

List of allowed publishers.

codeSigningPolicies :: CodeSigningPolicies

The code signing policy controls the validation failure action for signature mismatch or expiry.

lastModified :: Text

The date and time that the Code signing configuration was last modified, in ISO-8601 format (YYYY-MM-DDThh:mm:ss.sTZD).

newCodeSigningConfig

Arguments

:: Text
$sel:codeSigningConfigId:CodeSigningConfig'

-> Text
$sel:codeSigningConfigArn:CodeSigningConfig'

-> AllowedPublishers
$sel:allowedPublishers:CodeSigningConfig'

-> CodeSigningPolicies
$sel:codeSigningPolicies:CodeSigningConfig'

-> Text
$sel:lastModified:CodeSigningConfig'

-> CodeSigningConfig

Create a value of CodeSigningConfig with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:description:CodeSigningConfig', codeSigningConfig_description - Code signing configuration description.

$sel:codeSigningConfigId:CodeSigningConfig', codeSigningConfig_codeSigningConfigId - Unique identifer for the Code signing configuration.

$sel:codeSigningConfigArn:CodeSigningConfig', codeSigningConfig_codeSigningConfigArn - The Amazon Resource Name (ARN) of the Code signing configuration.

$sel:allowedPublishers:CodeSigningConfig', codeSigningConfig_allowedPublishers - List of allowed publishers.

$sel:codeSigningPolicies:CodeSigningConfig', codeSigningConfig_codeSigningPolicies - The code signing policy controls the validation failure action for signature mismatch or expiry.

$sel:lastModified:CodeSigningConfig', codeSigningConfig_lastModified - The date and time that the Code signing configuration was last modified, in ISO-8601 format (YYYY-MM-DDThh:mm:ss.sTZD).

codeSigningConfig_description :: Lens' CodeSigningConfig (Maybe Text)

Code signing configuration description.

codeSigningConfig_codeSigningConfigId :: Lens' CodeSigningConfig Text

Unique identifer for the Code signing configuration.

codeSigningConfig_codeSigningConfigArn :: Lens' CodeSigningConfig Text

The Amazon Resource Name (ARN) of the Code signing configuration.

codeSigningConfig_allowedPublishers :: Lens' CodeSigningConfig AllowedPublishers

List of allowed publishers.

codeSigningConfig_codeSigningPolicies :: Lens' CodeSigningConfig CodeSigningPolicies

The code signing policy controls the validation failure action for signature mismatch or expiry.

codeSigningConfig_lastModified :: Lens' CodeSigningConfig Text

The date and time that the Code signing configuration was last modified, in ISO-8601 format (YYYY-MM-DDThh:mm:ss.sTZD).