Amazonka.Lambda.UpdateFunctionConfiguration

Description

Modify the version-specific settings of a Lambda function.

When you update a function, Lambda provisions an instance of the function and its supporting resources. If your function connects to a VPC, this process can take a minute. During this time, you can't modify the function, but you can still invoke it. The LastUpdateStatus, LastUpdateStatusReason, and LastUpdateStatusReasonCode fields in the response from GetFunctionConfiguration indicate when the update is complete and the function is processing events with the new configuration. For more information, see Function States.

These settings can vary between versions of a function and are locked when you publish a version. You can't modify the configuration of a published version, only the unpublished version.

To configure function concurrency, use PutFunctionConcurrency. To grant invoke permissions to an account or Amazon Web Services service, use AddPermission.

Creating a Request

data UpdateFunctionConfiguration

See: newUpdateFunctionConfiguration smart constructor.

Constructors

UpdateFunctionConfiguration'

Fields

memorySize :: Maybe Natural

The amount of memory available to the function at runtime. Increasing the function memory also increases its CPU allocation. The default value is 128 MB. The value can be any multiple of 1 MB.

runtime :: Maybe Runtime

The identifier of the function's runtime.

kmsKeyArn :: Maybe Text

The ARN of the Amazon Web Services Key Management Service (KMS) key that's used to encrypt your function's environment variables. If it's not provided, Lambda uses a default service key.

fileSystemConfigs :: Maybe [FileSystemConfig]

Connection settings for an Amazon EFS file system.

environment :: Maybe Environment

Environment variables that are accessible from function code during execution.

imageConfig :: Maybe ImageConfig

Container image configuration values that override the values in the container image Docker file.

deadLetterConfig :: Maybe DeadLetterConfig

A dead letter queue configuration that specifies the queue or topic where Lambda sends asynchronous events when they fail processing. For more information, see Dead Letter Queues.

role' :: Maybe Text

The Amazon Resource Name (ARN) of the function's execution role.

vpcConfig :: Maybe VpcConfig

For network connectivity to Amazon Web Services resources in a VPC, specify a list of security groups and subnets in the VPC. When you connect a function to a VPC, it can only access resources and the internet through that VPC. For more information, see VPC Settings.

layers :: Maybe [Text]

A list of function layers to add to the function's execution environment. Specify each layer by its ARN, including the version.

handler :: Maybe Text

The name of the method within your code that Lambda calls to execute your function. The format includes the file name. It can also include namespaces and other qualifiers, depending on the runtime. For more information, see Programming Model.

timeout :: Maybe Natural

The amount of time that Lambda allows a function to run before stopping it. The default is 3 seconds. The maximum allowed value is 900 seconds. For additional information, see Lambda execution environment.

tracingConfig :: Maybe TracingConfig

Set Mode to Active to sample and trace a subset of incoming requests with X-Ray.

description :: Maybe Text

A description of the function.

revisionId :: Maybe Text

Only update the function if the revision ID matches the ID that's specified. Use this option to avoid modifying a function that has changed since you last read it.

functionName :: Text

The name of the Lambda function.

Name formats

Function name - my-function.

Function ARN - arn:aws:lambda:us-west-2:123456789012:function:my-function.

Partial ARN - 123456789012:function:my-function.

The length constraint applies only to the full ARN. If you specify only the function name, it is limited to 64 characters in length.

newUpdateFunctionConfiguration

Arguments

:: Text
$sel:functionName:UpdateFunctionConfiguration'

-> UpdateFunctionConfiguration

Create a value of UpdateFunctionConfiguration with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:memorySize:UpdateFunctionConfiguration', updateFunctionConfiguration_memorySize - The amount of memory available to the function at runtime. Increasing the function memory also increases its CPU allocation. The default value is 128 MB. The value can be any multiple of 1 MB.

$sel:runtime:UpdateFunctionConfiguration', updateFunctionConfiguration_runtime - The identifier of the function's runtime.

$sel:kmsKeyArn:UpdateFunctionConfiguration', updateFunctionConfiguration_kmsKeyArn - The ARN of the Amazon Web Services Key Management Service (KMS) key that's used to encrypt your function's environment variables. If it's not provided, Lambda uses a default service key.

$sel:fileSystemConfigs:UpdateFunctionConfiguration', updateFunctionConfiguration_fileSystemConfigs - Connection settings for an Amazon EFS file system.

$sel:environment:UpdateFunctionConfiguration', updateFunctionConfiguration_environment - Environment variables that are accessible from function code during execution.

$sel:imageConfig:UpdateFunctionConfiguration', updateFunctionConfiguration_imageConfig - Container image configuration values that override the values in the container image Docker file.

$sel:deadLetterConfig:UpdateFunctionConfiguration', updateFunctionConfiguration_deadLetterConfig - A dead letter queue configuration that specifies the queue or topic where Lambda sends asynchronous events when they fail processing. For more information, see Dead Letter Queues.

$sel:role':UpdateFunctionConfiguration', updateFunctionConfiguration_role - The Amazon Resource Name (ARN) of the function's execution role.

$sel:vpcConfig:UpdateFunctionConfiguration', updateFunctionConfiguration_vpcConfig - For network connectivity to Amazon Web Services resources in a VPC, specify a list of security groups and subnets in the VPC. When you connect a function to a VPC, it can only access resources and the internet through that VPC. For more information, see VPC Settings.

$sel:layers:UpdateFunctionConfiguration', updateFunctionConfiguration_layers - A list of function layers to add to the function's execution environment. Specify each layer by its ARN, including the version.

$sel:handler:UpdateFunctionConfiguration', updateFunctionConfiguration_handler - The name of the method within your code that Lambda calls to execute your function. The format includes the file name. It can also include namespaces and other qualifiers, depending on the runtime. For more information, see Programming Model.

$sel:timeout:UpdateFunctionConfiguration', updateFunctionConfiguration_timeout - The amount of time that Lambda allows a function to run before stopping it. The default is 3 seconds. The maximum allowed value is 900 seconds. For additional information, see Lambda execution environment.

$sel:tracingConfig:UpdateFunctionConfiguration', updateFunctionConfiguration_tracingConfig - Set Mode to Active to sample and trace a subset of incoming requests with X-Ray.

$sel:description:UpdateFunctionConfiguration', updateFunctionConfiguration_description - A description of the function.

$sel:revisionId:UpdateFunctionConfiguration', updateFunctionConfiguration_revisionId - Only update the function if the revision ID matches the ID that's specified. Use this option to avoid modifying a function that has changed since you last read it.

$sel:functionName:UpdateFunctionConfiguration', updateFunctionConfiguration_functionName - The name of the Lambda function.

Name formats

Function name - my-function.

Function ARN - arn:aws:lambda:us-west-2:123456789012:function:my-function.

Partial ARN - 123456789012:function:my-function.

The length constraint applies only to the full ARN. If you specify only the function name, it is limited to 64 characters in length.

Request Lenses

updateFunctionConfiguration_memorySize :: Lens' UpdateFunctionConfiguration (Maybe Natural)

The amount of memory available to the function at runtime. Increasing the function memory also increases its CPU allocation. The default value is 128 MB. The value can be any multiple of 1 MB.

updateFunctionConfiguration_runtime :: Lens' UpdateFunctionConfiguration (Maybe Runtime)

The identifier of the function's runtime.

updateFunctionConfiguration_kmsKeyArn :: Lens' UpdateFunctionConfiguration (Maybe Text)

The ARN of the Amazon Web Services Key Management Service (KMS) key that's used to encrypt your function's environment variables. If it's not provided, Lambda uses a default service key.

updateFunctionConfiguration_fileSystemConfigs :: Lens' UpdateFunctionConfiguration (Maybe [FileSystemConfig])

Connection settings for an Amazon EFS file system.

updateFunctionConfiguration_environment :: Lens' UpdateFunctionConfiguration (Maybe Environment)

Environment variables that are accessible from function code during execution.

updateFunctionConfiguration_imageConfig :: Lens' UpdateFunctionConfiguration (Maybe ImageConfig)

Container image configuration values that override the values in the container image Docker file.

updateFunctionConfiguration_deadLetterConfig :: Lens' UpdateFunctionConfiguration (Maybe DeadLetterConfig)

A dead letter queue configuration that specifies the queue or topic where Lambda sends asynchronous events when they fail processing. For more information, see Dead Letter Queues.

updateFunctionConfiguration_role :: Lens' UpdateFunctionConfiguration (Maybe Text)

The Amazon Resource Name (ARN) of the function's execution role.

updateFunctionConfiguration_vpcConfig :: Lens' UpdateFunctionConfiguration (Maybe VpcConfig)

For network connectivity to Amazon Web Services resources in a VPC, specify a list of security groups and subnets in the VPC. When you connect a function to a VPC, it can only access resources and the internet through that VPC. For more information, see VPC Settings.

updateFunctionConfiguration_layers :: Lens' UpdateFunctionConfiguration (Maybe [Text])

A list of function layers to add to the function's execution environment. Specify each layer by its ARN, including the version.

updateFunctionConfiguration_handler :: Lens' UpdateFunctionConfiguration (Maybe Text)

The name of the method within your code that Lambda calls to execute your function. The format includes the file name. It can also include namespaces and other qualifiers, depending on the runtime. For more information, see Programming Model.

updateFunctionConfiguration_timeout :: Lens' UpdateFunctionConfiguration (Maybe Natural)

The amount of time that Lambda allows a function to run before stopping it. The default is 3 seconds. The maximum allowed value is 900 seconds. For additional information, see Lambda execution environment.

updateFunctionConfiguration_tracingConfig :: Lens' UpdateFunctionConfiguration (Maybe TracingConfig)

Set Mode to Active to sample and trace a subset of incoming requests with X-Ray.

updateFunctionConfiguration_description :: Lens' UpdateFunctionConfiguration (Maybe Text)

A description of the function.

updateFunctionConfiguration_revisionId :: Lens' UpdateFunctionConfiguration (Maybe Text)

Only update the function if the revision ID matches the ID that's specified. Use this option to avoid modifying a function that has changed since you last read it.

updateFunctionConfiguration_functionName :: Lens' UpdateFunctionConfiguration Text

The name of the Lambda function.

Name formats

Function name - my-function.

Function ARN - arn:aws:lambda:us-west-2:123456789012:function:my-function.

Partial ARN - 123456789012:function:my-function.

The length constraint applies only to the full ARN. If you specify only the function name, it is limited to 64 characters in length.

Destructuring the Response

data FunctionConfiguration

Details about a function's configuration.

See: newFunctionConfiguration smart constructor.

Constructors

FunctionConfiguration'

Fields

memorySize :: Maybe Natural

The amount of memory available to the function at runtime.

runtime :: Maybe Runtime

The runtime environment for the Lambda function.

state :: Maybe State

The current state of the function. When the state is Inactive, you can reactivate the function by invoking it.

signingProfileVersionArn :: Maybe Text

The ARN of the signing profile version.

lastUpdateStatus :: Maybe LastUpdateStatus

The status of the last update that was performed on the function. This is first set to Successful after function creation completes.

functionArn :: Maybe Text

The function's Amazon Resource Name (ARN).

kmsKeyArn :: Maybe Text

The KMS key that's used to encrypt the function's environment variables. This key is only returned if you've configured a customer managed CMK.

packageType :: Maybe PackageType

The type of deployment package. Set to Image for container image and set Zip for .zip file archive.

fileSystemConfigs :: Maybe [FileSystemConfig]

Connection settings for an Amazon EFS file system.

environment :: Maybe EnvironmentResponse

The function's environment variables.

deadLetterConfig :: Maybe DeadLetterConfig

The function's dead letter queue.

architectures :: Maybe (NonEmpty Architecture)

The instruction set architecture that the function supports. Architecture is a string array with one of the valid values. The default architecture value is x86_64.

signingJobArn :: Maybe Text

The ARN of the signing job.

role' :: Maybe Text

The function's execution role.

vpcConfig :: Maybe VpcConfigResponse

The function's networking configuration.

version :: Maybe Text

The version of the Lambda function.

functionName :: Maybe Text

The name of the function.

layers :: Maybe [Layer]

The function's layers.

codeSize :: Maybe Integer

The size of the function's deployment package, in bytes.

handler :: Maybe Text

The function that Lambda calls to begin executing your function.

timeout :: Maybe Natural

The amount of time in seconds that Lambda allows a function to run before stopping it.

lastUpdateStatusReason :: Maybe Text

The reason for the last update that was performed on the function.

stateReason :: Maybe Text

The reason for the function's current state.

lastModified :: Maybe Text

The date and time that the function was last updated, in ISO-8601 format (YYYY-MM-DDThh:mm:ss.sTZD).

codeSha256 :: Maybe Text

The SHA256 hash of the function's deployment package.

tracingConfig :: Maybe TracingConfigResponse

The function's X-Ray tracing configuration.

stateReasonCode :: Maybe StateReasonCode

The reason code for the function's current state. When the code is Creating, you can't invoke or modify the function.

imageConfigResponse :: Maybe ImageConfigResponse

The function's image configuration values.

description :: Maybe Text

The function's description.

lastUpdateStatusReasonCode :: Maybe LastUpdateStatusReasonCode

The reason code for the last update that was performed on the function.

revisionId :: Maybe Text

The latest updated revision of the function or alias.

masterArn :: Maybe Text

For Lambda@Edge functions, the ARN of the master function.

newFunctionConfiguration :: FunctionConfiguration

Create a value of FunctionConfiguration with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:memorySize:FunctionConfiguration', functionConfiguration_memorySize - The amount of memory available to the function at runtime.

$sel:runtime:FunctionConfiguration', functionConfiguration_runtime - The runtime environment for the Lambda function.

$sel:state:FunctionConfiguration', functionConfiguration_state - The current state of the function. When the state is Inactive, you can reactivate the function by invoking it.

$sel:signingProfileVersionArn:FunctionConfiguration', functionConfiguration_signingProfileVersionArn - The ARN of the signing profile version.

$sel:lastUpdateStatus:FunctionConfiguration', functionConfiguration_lastUpdateStatus - The status of the last update that was performed on the function. This is first set to Successful after function creation completes.

$sel:functionArn:FunctionConfiguration', functionConfiguration_functionArn - The function's Amazon Resource Name (ARN).

$sel:kmsKeyArn:FunctionConfiguration', functionConfiguration_kmsKeyArn - The KMS key that's used to encrypt the function's environment variables. This key is only returned if you've configured a customer managed CMK.

$sel:packageType:FunctionConfiguration', functionConfiguration_packageType - The type of deployment package. Set to Image for container image and set Zip for .zip file archive.

$sel:fileSystemConfigs:FunctionConfiguration', functionConfiguration_fileSystemConfigs - Connection settings for an Amazon EFS file system.

$sel:environment:FunctionConfiguration', functionConfiguration_environment - The function's environment variables.

$sel:deadLetterConfig:FunctionConfiguration', functionConfiguration_deadLetterConfig - The function's dead letter queue.

$sel:architectures:FunctionConfiguration', functionConfiguration_architectures - The instruction set architecture that the function supports. Architecture is a string array with one of the valid values. The default architecture value is x86_64.

$sel:signingJobArn:FunctionConfiguration', functionConfiguration_signingJobArn - The ARN of the signing job.

$sel:role':FunctionConfiguration', functionConfiguration_role - The function's execution role.

$sel:vpcConfig:FunctionConfiguration', functionConfiguration_vpcConfig - The function's networking configuration.

$sel:version:FunctionConfiguration', functionConfiguration_version - The version of the Lambda function.

$sel:functionName:FunctionConfiguration', functionConfiguration_functionName - The name of the function.

$sel:layers:FunctionConfiguration', functionConfiguration_layers - The function's layers.

$sel:codeSize:FunctionConfiguration', functionConfiguration_codeSize - The size of the function's deployment package, in bytes.

$sel:handler:FunctionConfiguration', functionConfiguration_handler - The function that Lambda calls to begin executing your function.

$sel:timeout:FunctionConfiguration', functionConfiguration_timeout - The amount of time in seconds that Lambda allows a function to run before stopping it.

$sel:lastUpdateStatusReason:FunctionConfiguration', functionConfiguration_lastUpdateStatusReason - The reason for the last update that was performed on the function.

$sel:stateReason:FunctionConfiguration', functionConfiguration_stateReason - The reason for the function's current state.

$sel:lastModified:FunctionConfiguration', functionConfiguration_lastModified - The date and time that the function was last updated, in ISO-8601 format (YYYY-MM-DDThh:mm:ss.sTZD).

$sel:codeSha256:FunctionConfiguration', functionConfiguration_codeSha256 - The SHA256 hash of the function's deployment package.

$sel:tracingConfig:FunctionConfiguration', functionConfiguration_tracingConfig - The function's X-Ray tracing configuration.

$sel:stateReasonCode:FunctionConfiguration', functionConfiguration_stateReasonCode - The reason code for the function's current state. When the code is Creating, you can't invoke or modify the function.

$sel:imageConfigResponse:FunctionConfiguration', functionConfiguration_imageConfigResponse - The function's image configuration values.

$sel:description:FunctionConfiguration', functionConfiguration_description - The function's description.

$sel:lastUpdateStatusReasonCode:FunctionConfiguration', functionConfiguration_lastUpdateStatusReasonCode - The reason code for the last update that was performed on the function.

$sel:revisionId:FunctionConfiguration', functionConfiguration_revisionId - The latest updated revision of the function or alias.

$sel:masterArn:FunctionConfiguration', functionConfiguration_masterArn - For Lambda@Edge functions, the ARN of the master function.

Response Lenses

functionConfiguration_memorySize :: Lens' FunctionConfiguration (Maybe Natural)

The amount of memory available to the function at runtime.

functionConfiguration_runtime :: Lens' FunctionConfiguration (Maybe Runtime)

The runtime environment for the Lambda function.

functionConfiguration_state :: Lens' FunctionConfiguration (Maybe State)

The current state of the function. When the state is Inactive, you can reactivate the function by invoking it.

functionConfiguration_signingProfileVersionArn :: Lens' FunctionConfiguration (Maybe Text)

The ARN of the signing profile version.

functionConfiguration_lastUpdateStatus :: Lens' FunctionConfiguration (Maybe LastUpdateStatus)

The status of the last update that was performed on the function. This is first set to Successful after function creation completes.

functionConfiguration_functionArn :: Lens' FunctionConfiguration (Maybe Text)

The function's Amazon Resource Name (ARN).

functionConfiguration_kmsKeyArn :: Lens' FunctionConfiguration (Maybe Text)

The KMS key that's used to encrypt the function's environment variables. This key is only returned if you've configured a customer managed CMK.

functionConfiguration_packageType :: Lens' FunctionConfiguration (Maybe PackageType)

The type of deployment package. Set to Image for container image and set Zip for .zip file archive.

functionConfiguration_fileSystemConfigs :: Lens' FunctionConfiguration (Maybe [FileSystemConfig])

Connection settings for an Amazon EFS file system.

functionConfiguration_environment :: Lens' FunctionConfiguration (Maybe EnvironmentResponse)

The function's environment variables.

functionConfiguration_deadLetterConfig :: Lens' FunctionConfiguration (Maybe DeadLetterConfig)

The function's dead letter queue.

functionConfiguration_architectures :: Lens' FunctionConfiguration (Maybe (NonEmpty Architecture))

The instruction set architecture that the function supports. Architecture is a string array with one of the valid values. The default architecture value is x86_64.

functionConfiguration_signingJobArn :: Lens' FunctionConfiguration (Maybe Text)

The ARN of the signing job.

functionConfiguration_role :: Lens' FunctionConfiguration (Maybe Text)

The function's execution role.

functionConfiguration_vpcConfig :: Lens' FunctionConfiguration (Maybe VpcConfigResponse)

The function's networking configuration.

functionConfiguration_version :: Lens' FunctionConfiguration (Maybe Text)

The version of the Lambda function.

functionConfiguration_functionName :: Lens' FunctionConfiguration (Maybe Text)

The name of the function.

functionConfiguration_layers :: Lens' FunctionConfiguration (Maybe [Layer])

The function's layers.

functionConfiguration_codeSize :: Lens' FunctionConfiguration (Maybe Integer)

The size of the function's deployment package, in bytes.

functionConfiguration_handler :: Lens' FunctionConfiguration (Maybe Text)

The function that Lambda calls to begin executing your function.

functionConfiguration_timeout :: Lens' FunctionConfiguration (Maybe Natural)

The amount of time in seconds that Lambda allows a function to run before stopping it.

functionConfiguration_lastUpdateStatusReason :: Lens' FunctionConfiguration (Maybe Text)

The reason for the last update that was performed on the function.

functionConfiguration_stateReason :: Lens' FunctionConfiguration (Maybe Text)

The reason for the function's current state.

functionConfiguration_lastModified :: Lens' FunctionConfiguration (Maybe Text)

The date and time that the function was last updated, in ISO-8601 format (YYYY-MM-DDThh:mm:ss.sTZD).

functionConfiguration_codeSha256 :: Lens' FunctionConfiguration (Maybe Text)

The SHA256 hash of the function's deployment package.

functionConfiguration_tracingConfig :: Lens' FunctionConfiguration (Maybe TracingConfigResponse)

The function's X-Ray tracing configuration.

functionConfiguration_stateReasonCode :: Lens' FunctionConfiguration (Maybe StateReasonCode)

The reason code for the function's current state. When the code is Creating, you can't invoke or modify the function.

functionConfiguration_imageConfigResponse :: Lens' FunctionConfiguration (Maybe ImageConfigResponse)

The function's image configuration values.

functionConfiguration_description :: Lens' FunctionConfiguration (Maybe Text)

The function's description.

functionConfiguration_lastUpdateStatusReasonCode :: Lens' FunctionConfiguration (Maybe LastUpdateStatusReasonCode)

The reason code for the last update that was performed on the function.

functionConfiguration_revisionId :: Lens' FunctionConfiguration (Maybe Text)

The latest updated revision of the function or alias.

functionConfiguration_masterArn :: Lens' FunctionConfiguration (Maybe Text)

For Lambda@Edge functions, the ARN of the master function.