Amazonka.LexModels.DeleteBotChannelAssociation

Description

Deletes the association between an Amazon Lex bot and a messaging platform.

This operation requires permission for the lex:DeleteBotChannelAssociation action.

Creating a Request

data DeleteBotChannelAssociation

See: newDeleteBotChannelAssociation smart constructor.

Constructors

DeleteBotChannelAssociation'

Fields

name :: Text

The name of the association. The name is case sensitive.

botName :: Text

The name of the Amazon Lex bot.

botAlias :: Text

An alias that points to the specific version of the Amazon Lex bot to which this association is being made.

newDeleteBotChannelAssociation

Arguments

:: Text
$sel:name:DeleteBotChannelAssociation'

-> Text
$sel:botName:DeleteBotChannelAssociation'

-> Text
$sel:botAlias:DeleteBotChannelAssociation'

-> DeleteBotChannelAssociation

Create a value of DeleteBotChannelAssociation with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:name:DeleteBotChannelAssociation', deleteBotChannelAssociation_name - The name of the association. The name is case sensitive.

$sel:botName:DeleteBotChannelAssociation', deleteBotChannelAssociation_botName - The name of the Amazon Lex bot.

$sel:botAlias:DeleteBotChannelAssociation', deleteBotChannelAssociation_botAlias - An alias that points to the specific version of the Amazon Lex bot to which this association is being made.

Request Lenses

deleteBotChannelAssociation_name :: Lens' DeleteBotChannelAssociation Text

The name of the association. The name is case sensitive.

deleteBotChannelAssociation_botName :: Lens' DeleteBotChannelAssociation Text

The name of the Amazon Lex bot.

deleteBotChannelAssociation_botAlias :: Lens' DeleteBotChannelAssociation Text

An alias that points to the specific version of the Amazon Lex bot to which this association is being made.

Destructuring the Response

data DeleteBotChannelAssociationResponse

See: newDeleteBotChannelAssociationResponse smart constructor.

Constructors

DeleteBotChannelAssociationResponse'

newDeleteBotChannelAssociationResponse :: DeleteBotChannelAssociationResponse

Create a value of DeleteBotChannelAssociationResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.