Amazonka.LexModels.GetBot

Description

Returns metadata information for a specific bot. You must provide the bot name and the bot version or alias.

This operation requires permissions for the lex:GetBot action.

Creating a Request

data GetBot

See: newGetBot smart constructor.

Constructors

GetBot'

Fields

name :: Text

The name of the bot. The name is case sensitive.

versionOrAlias :: Text

The version or alias of the bot.

newGetBot

Arguments

:: Text
$sel:name:GetBot'

-> Text
$sel:versionOrAlias:GetBot'

-> GetBot

Create a value of GetBot with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:name:GetBot', getBot_name - The name of the bot. The name is case sensitive.

$sel:versionOrAlias:GetBot', getBot_versionOrAlias - The version or alias of the bot.

Request Lenses

getBot_name :: Lens' GetBot Text

The name of the bot. The name is case sensitive.

getBot_versionOrAlias :: Lens' GetBot Text

The version or alias of the bot.

Destructuring the Response

data GetBotResponse

See: newGetBotResponse smart constructor.

Constructors

GetBotResponse'

Fields

failureReason :: Maybe Text

If status is FAILED, Amazon Lex explains why it failed to build the bot.

status :: Maybe LexStatus

The status of the bot.

When the status is BUILDING Amazon Lex is building the bot for testing and use.

If the status of the bot is READY_BASIC_TESTING, you can test the bot using the exact utterances specified in the bot's intents. When the bot is ready for full testing or to run, the status is READY.

If there was a problem with building the bot, the status is FAILED and the failureReason field explains why the bot did not build.

If the bot was saved but not built, the status is NOT_BUILT.

abortStatement :: Maybe Statement

The message that Amazon Lex returns when the user elects to end the conversation without completing it. For more information, see PutBot.

intents :: Maybe [Intent]

An array of intent objects. For more information, see PutBot.

checksum :: Maybe Text

Checksum of the bot used to identify a specific revision of the bot's $LATEST version.

enableModelImprovements :: Maybe Bool

Indicates whether the bot uses accuracy improvements. true indicates that the bot is using the improvements, otherwise, false.

nluIntentConfidenceThreshold :: Maybe Double

The score that determines where Amazon Lex inserts the AMAZON.FallbackIntent, AMAZON.KendraSearchIntent, or both when returning alternative intents in a PostContent or PostText response. AMAZON.FallbackIntent is inserted if the confidence score for all intents is below this value. AMAZON.KendraSearchIntent is only inserted if it is configured for the bot.

detectSentiment :: Maybe Bool

Indicates whether user utterances should be sent to Amazon Comprehend for sentiment analysis.

locale :: Maybe Locale

The target locale for the bot.

createdDate :: Maybe POSIX

The date that the bot was created.

name :: Maybe Text

The name of the bot.

version :: Maybe Text

The version of the bot. For a new bot, the version is always $LATEST.

idleSessionTTLInSeconds :: Maybe Natural

The maximum time in seconds that Amazon Lex retains the data gathered in a conversation. For more information, see PutBot.

clarificationPrompt :: Maybe Prompt

The message Amazon Lex uses when it doesn't understand the user's request. For more information, see PutBot.

voiceId :: Maybe Text

The Amazon Polly voice ID that Amazon Lex uses for voice interaction with the user. For more information, see PutBot.

lastUpdatedDate :: Maybe POSIX

The date that the bot was updated. When you create a resource, the creation date and last updated date are the same.

childDirected :: Maybe Bool

For each Amazon Lex bot created with the Amazon Lex Model Building Service, you must specify whether your use of Amazon Lex is related to a website, program, or other application that is directed or targeted, in whole or in part, to children under age 13 and subject to the Children's Online Privacy Protection Act (COPPA) by specifying true or false in the childDirected field. By specifying true in the childDirected field, you confirm that your use of Amazon Lex is related to a website, program, or other application that is directed or targeted, in whole or in part, to children under age 13 and subject to COPPA. By specifying false in the childDirected field, you confirm that your use of Amazon Lex is not related to a website, program, or other application that is directed or targeted, in whole or in part, to children under age 13 and subject to COPPA. You may not specify a default value for the childDirected field that does not accurately reflect whether your use of Amazon Lex is related to a website, program, or other application that is directed or targeted, in whole or in part, to children under age 13 and subject to COPPA.

If your use of Amazon Lex relates to a website, program, or other application that is directed in whole or in part, to children under age 13, you must obtain any required verifiable parental consent under COPPA. For information regarding the use of Amazon Lex in connection with websites, programs, or other applications that are directed or targeted, in whole or in part, to children under age 13, see the Amazon Lex FAQ.

description :: Maybe Text

A description of the bot.

httpStatus :: Int

The response's http status code.

newGetBotResponse

Arguments

:: Int
$sel:httpStatus:GetBotResponse'

-> GetBotResponse

Create a value of GetBotResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:failureReason:GetBotResponse', getBotResponse_failureReason - If status is FAILED, Amazon Lex explains why it failed to build the bot.

$sel:status:GetBotResponse', getBotResponse_status - The status of the bot.

When the status is BUILDING Amazon Lex is building the bot for testing and use.

If the status of the bot is READY_BASIC_TESTING, you can test the bot using the exact utterances specified in the bot's intents. When the bot is ready for full testing or to run, the status is READY.

If there was a problem with building the bot, the status is FAILED and the failureReason field explains why the bot did not build.

If the bot was saved but not built, the status is NOT_BUILT.

$sel:abortStatement:GetBotResponse', getBotResponse_abortStatement - The message that Amazon Lex returns when the user elects to end the conversation without completing it. For more information, see PutBot.

$sel:intents:GetBotResponse', getBotResponse_intents - An array of intent objects. For more information, see PutBot.

$sel:checksum:GetBotResponse', getBotResponse_checksum - Checksum of the bot used to identify a specific revision of the bot's $LATEST version.

$sel:enableModelImprovements:GetBotResponse', getBotResponse_enableModelImprovements - Indicates whether the bot uses accuracy improvements. true indicates that the bot is using the improvements, otherwise, false.

$sel:nluIntentConfidenceThreshold:GetBotResponse', getBotResponse_nluIntentConfidenceThreshold - The score that determines where Amazon Lex inserts the AMAZON.FallbackIntent, AMAZON.KendraSearchIntent, or both when returning alternative intents in a PostContent or PostText response. AMAZON.FallbackIntent is inserted if the confidence score for all intents is below this value. AMAZON.KendraSearchIntent is only inserted if it is configured for the bot.

$sel:detectSentiment:GetBotResponse', getBotResponse_detectSentiment - Indicates whether user utterances should be sent to Amazon Comprehend for sentiment analysis.

$sel:locale:GetBotResponse', getBotResponse_locale - The target locale for the bot.

$sel:createdDate:GetBotResponse', getBotResponse_createdDate - The date that the bot was created.

$sel:name:GetBot', getBotResponse_name - The name of the bot.

$sel:version:GetBotResponse', getBotResponse_version - The version of the bot. For a new bot, the version is always $LATEST.

$sel:idleSessionTTLInSeconds:GetBotResponse', getBotResponse_idleSessionTTLInSeconds - The maximum time in seconds that Amazon Lex retains the data gathered in a conversation. For more information, see PutBot.

$sel:clarificationPrompt:GetBotResponse', getBotResponse_clarificationPrompt - The message Amazon Lex uses when it doesn't understand the user's request. For more information, see PutBot.

$sel:voiceId:GetBotResponse', getBotResponse_voiceId - The Amazon Polly voice ID that Amazon Lex uses for voice interaction with the user. For more information, see PutBot.

$sel:lastUpdatedDate:GetBotResponse', getBotResponse_lastUpdatedDate - The date that the bot was updated. When you create a resource, the creation date and last updated date are the same.

$sel:childDirected:GetBotResponse', getBotResponse_childDirected - For each Amazon Lex bot created with the Amazon Lex Model Building Service, you must specify whether your use of Amazon Lex is related to a website, program, or other application that is directed or targeted, in whole or in part, to children under age 13 and subject to the Children's Online Privacy Protection Act (COPPA) by specifying true or false in the childDirected field. By specifying true in the childDirected field, you confirm that your use of Amazon Lex is related to a website, program, or other application that is directed or targeted, in whole or in part, to children under age 13 and subject to COPPA. By specifying false in the childDirected field, you confirm that your use of Amazon Lex is not related to a website, program, or other application that is directed or targeted, in whole or in part, to children under age 13 and subject to COPPA. You may not specify a default value for the childDirected field that does not accurately reflect whether your use of Amazon Lex is related to a website, program, or other application that is directed or targeted, in whole or in part, to children under age 13 and subject to COPPA.

If your use of Amazon Lex relates to a website, program, or other application that is directed in whole or in part, to children under age 13, you must obtain any required verifiable parental consent under COPPA. For information regarding the use of Amazon Lex in connection with websites, programs, or other applications that are directed or targeted, in whole or in part, to children under age 13, see the Amazon Lex FAQ.

$sel:description:GetBotResponse', getBotResponse_description - A description of the bot.

$sel:httpStatus:GetBotResponse', getBotResponse_httpStatus - The response's http status code.

Response Lenses

getBotResponse_failureReason :: Lens' GetBotResponse (Maybe Text)

If status is FAILED, Amazon Lex explains why it failed to build the bot.

getBotResponse_status :: Lens' GetBotResponse (Maybe LexStatus)

The status of the bot.

When the status is BUILDING Amazon Lex is building the bot for testing and use.

If the status of the bot is READY_BASIC_TESTING, you can test the bot using the exact utterances specified in the bot's intents. When the bot is ready for full testing or to run, the status is READY.

If there was a problem with building the bot, the status is FAILED and the failureReason field explains why the bot did not build.

If the bot was saved but not built, the status is NOT_BUILT.

getBotResponse_abortStatement :: Lens' GetBotResponse (Maybe Statement)

The message that Amazon Lex returns when the user elects to end the conversation without completing it. For more information, see PutBot.

getBotResponse_intents :: Lens' GetBotResponse (Maybe [Intent])

An array of intent objects. For more information, see PutBot.

getBotResponse_checksum :: Lens' GetBotResponse (Maybe Text)

Checksum of the bot used to identify a specific revision of the bot's $LATEST version.

getBotResponse_enableModelImprovements :: Lens' GetBotResponse (Maybe Bool)

Indicates whether the bot uses accuracy improvements. true indicates that the bot is using the improvements, otherwise, false.

getBotResponse_nluIntentConfidenceThreshold :: Lens' GetBotResponse (Maybe Double)

The score that determines where Amazon Lex inserts the AMAZON.FallbackIntent, AMAZON.KendraSearchIntent, or both when returning alternative intents in a PostContent or PostText response. AMAZON.FallbackIntent is inserted if the confidence score for all intents is below this value. AMAZON.KendraSearchIntent is only inserted if it is configured for the bot.

getBotResponse_detectSentiment :: Lens' GetBotResponse (Maybe Bool)

Indicates whether user utterances should be sent to Amazon Comprehend for sentiment analysis.

getBotResponse_locale :: Lens' GetBotResponse (Maybe Locale)

The target locale for the bot.

getBotResponse_createdDate :: Lens' GetBotResponse (Maybe UTCTime)

The date that the bot was created.

getBotResponse_name :: Lens' GetBotResponse (Maybe Text)

The name of the bot.

getBotResponse_version :: Lens' GetBotResponse (Maybe Text)

The version of the bot. For a new bot, the version is always $LATEST.

getBotResponse_idleSessionTTLInSeconds :: Lens' GetBotResponse (Maybe Natural)

The maximum time in seconds that Amazon Lex retains the data gathered in a conversation. For more information, see PutBot.

getBotResponse_clarificationPrompt :: Lens' GetBotResponse (Maybe Prompt)

The message Amazon Lex uses when it doesn't understand the user's request. For more information, see PutBot.

getBotResponse_voiceId :: Lens' GetBotResponse (Maybe Text)

The Amazon Polly voice ID that Amazon Lex uses for voice interaction with the user. For more information, see PutBot.

getBotResponse_lastUpdatedDate :: Lens' GetBotResponse (Maybe UTCTime)

The date that the bot was updated. When you create a resource, the creation date and last updated date are the same.

getBotResponse_childDirected :: Lens' GetBotResponse (Maybe Bool)

For each Amazon Lex bot created with the Amazon Lex Model Building Service, you must specify whether your use of Amazon Lex is related to a website, program, or other application that is directed or targeted, in whole or in part, to children under age 13 and subject to the Children's Online Privacy Protection Act (COPPA) by specifying true or false in the childDirected field. By specifying true in the childDirected field, you confirm that your use of Amazon Lex is related to a website, program, or other application that is directed or targeted, in whole or in part, to children under age 13 and subject to COPPA. By specifying false in the childDirected field, you confirm that your use of Amazon Lex is not related to a website, program, or other application that is directed or targeted, in whole or in part, to children under age 13 and subject to COPPA. You may not specify a default value for the childDirected field that does not accurately reflect whether your use of Amazon Lex is related to a website, program, or other application that is directed or targeted, in whole or in part, to children under age 13 and subject to COPPA.

If your use of Amazon Lex relates to a website, program, or other application that is directed in whole or in part, to children under age 13, you must obtain any required verifiable parental consent under COPPA. For information regarding the use of Amazon Lex in connection with websites, programs, or other applications that are directed or targeted, in whole or in part, to children under age 13, see the Amazon Lex FAQ.

getBotResponse_description :: Lens' GetBotResponse (Maybe Text)

A description of the bot.

getBotResponse_httpStatus :: Lens' GetBotResponse Int

The response's http status code.