Amazonka.LexModels.GetBuiltinIntents

Description

Gets a list of built-in intents that meet the specified criteria.

This operation requires permission for the lex:GetBuiltinIntents action.

This operation returns paginated results.

Creating a Request

data GetBuiltinIntents

See: newGetBuiltinIntents smart constructor.

Constructors

GetBuiltinIntents'

Fields

locale :: Maybe Locale

A list of locales that the intent supports.

nextToken :: Maybe Text

A pagination token that fetches the next page of intents. If this API call is truncated, Amazon Lex returns a pagination token in the response. To fetch the next page of intents, use the pagination token in the next request.

signatureContains :: Maybe Text

Substring to match in built-in intent signatures. An intent will be returned if any part of its signature matches the substring. For example, "xyz" matches both "xyzabc" and "abcxyz." To find the signature for an intent, see Standard Built-in Intents in the Alexa Skills Kit.

maxResults :: Maybe Natural

The maximum number of intents to return in the response. The default is 10.

newGetBuiltinIntents :: GetBuiltinIntents

Create a value of GetBuiltinIntents with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:locale:GetBuiltinIntents', getBuiltinIntents_locale - A list of locales that the intent supports.

$sel:nextToken:GetBuiltinIntents', getBuiltinIntents_nextToken - A pagination token that fetches the next page of intents. If this API call is truncated, Amazon Lex returns a pagination token in the response. To fetch the next page of intents, use the pagination token in the next request.

$sel:signatureContains:GetBuiltinIntents', getBuiltinIntents_signatureContains - Substring to match in built-in intent signatures. An intent will be returned if any part of its signature matches the substring. For example, "xyz" matches both "xyzabc" and "abcxyz." To find the signature for an intent, see Standard Built-in Intents in the Alexa Skills Kit.

$sel:maxResults:GetBuiltinIntents', getBuiltinIntents_maxResults - The maximum number of intents to return in the response. The default is 10.

Request Lenses

getBuiltinIntents_locale :: Lens' GetBuiltinIntents (Maybe Locale)

A list of locales that the intent supports.

getBuiltinIntents_nextToken :: Lens' GetBuiltinIntents (Maybe Text)

A pagination token that fetches the next page of intents. If this API call is truncated, Amazon Lex returns a pagination token in the response. To fetch the next page of intents, use the pagination token in the next request.

getBuiltinIntents_signatureContains :: Lens' GetBuiltinIntents (Maybe Text)

Substring to match in built-in intent signatures. An intent will be returned if any part of its signature matches the substring. For example, "xyz" matches both "xyzabc" and "abcxyz." To find the signature for an intent, see Standard Built-in Intents in the Alexa Skills Kit.

getBuiltinIntents_maxResults :: Lens' GetBuiltinIntents (Maybe Natural)

The maximum number of intents to return in the response. The default is 10.

Destructuring the Response

data GetBuiltinIntentsResponse

See: newGetBuiltinIntentsResponse smart constructor.

Constructors

GetBuiltinIntentsResponse'

Fields

intents :: Maybe [BuiltinIntentMetadata]

An array of builtinIntentMetadata objects, one for each intent in the response.

nextToken :: Maybe Text

A pagination token that fetches the next page of intents. If the response to this API call is truncated, Amazon Lex returns a pagination token in the response. To fetch the next page of intents, specify the pagination token in the next request.

httpStatus :: Int

The response's http status code.

newGetBuiltinIntentsResponse

Arguments

:: Int
$sel:httpStatus:GetBuiltinIntentsResponse'

-> GetBuiltinIntentsResponse

Create a value of GetBuiltinIntentsResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:intents:GetBuiltinIntentsResponse', getBuiltinIntentsResponse_intents - An array of builtinIntentMetadata objects, one for each intent in the response.

$sel:nextToken:GetBuiltinIntents', getBuiltinIntentsResponse_nextToken - A pagination token that fetches the next page of intents. If the response to this API call is truncated, Amazon Lex returns a pagination token in the response. To fetch the next page of intents, specify the pagination token in the next request.

$sel:httpStatus:GetBuiltinIntentsResponse', getBuiltinIntentsResponse_httpStatus - The response's http status code.

Response Lenses

getBuiltinIntentsResponse_intents :: Lens' GetBuiltinIntentsResponse (Maybe [BuiltinIntentMetadata])

An array of builtinIntentMetadata objects, one for each intent in the response.

getBuiltinIntentsResponse_nextToken :: Lens' GetBuiltinIntentsResponse (Maybe Text)

A pagination token that fetches the next page of intents. If the response to this API call is truncated, Amazon Lex returns a pagination token in the response. To fetch the next page of intents, specify the pagination token in the next request.

getBuiltinIntentsResponse_httpStatus :: Lens' GetBuiltinIntentsResponse Int

The response's http status code.