Amazonka.LexModels.GetIntentVersions

Description

Gets information about all of the versions of an intent.

The GetIntentVersions operation returns an IntentMetadata object for each version of an intent. For example, if an intent has three numbered versions, the GetIntentVersions operation returns four IntentMetadata objects in the response, one for each numbered version and one for the $LATEST version.

The GetIntentVersions operation always returns at least one version, the $LATEST version.

This operation requires permissions for the lex:GetIntentVersions action.

This operation returns paginated results.

Creating a Request

data GetIntentVersions

See: newGetIntentVersions smart constructor.

Constructors

GetIntentVersions'

Fields

nextToken :: Maybe Text

A pagination token for fetching the next page of intent versions. If the response to this call is truncated, Amazon Lex returns a pagination token in the response. To fetch the next page of versions, specify the pagination token in the next request.

maxResults :: Maybe Natural

The maximum number of intent versions to return in the response. The default is 10.

name :: Text

The name of the intent for which versions should be returned.

newGetIntentVersions

Arguments

:: Text
$sel:name:GetIntentVersions'

-> GetIntentVersions

Create a value of GetIntentVersions with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:nextToken:GetIntentVersions', getIntentVersions_nextToken - A pagination token for fetching the next page of intent versions. If the response to this call is truncated, Amazon Lex returns a pagination token in the response. To fetch the next page of versions, specify the pagination token in the next request.

$sel:maxResults:GetIntentVersions', getIntentVersions_maxResults - The maximum number of intent versions to return in the response. The default is 10.

$sel:name:GetIntentVersions', getIntentVersions_name - The name of the intent for which versions should be returned.

Request Lenses

getIntentVersions_nextToken :: Lens' GetIntentVersions (Maybe Text)

A pagination token for fetching the next page of intent versions. If the response to this call is truncated, Amazon Lex returns a pagination token in the response. To fetch the next page of versions, specify the pagination token in the next request.

getIntentVersions_maxResults :: Lens' GetIntentVersions (Maybe Natural)

The maximum number of intent versions to return in the response. The default is 10.

getIntentVersions_name :: Lens' GetIntentVersions Text

The name of the intent for which versions should be returned.

Destructuring the Response

data GetIntentVersionsResponse

See: newGetIntentVersionsResponse smart constructor.

Constructors

GetIntentVersionsResponse'

Fields

intents :: Maybe [IntentMetadata]

An array of IntentMetadata objects, one for each numbered version of the intent plus one for the $LATEST version.

nextToken :: Maybe Text

A pagination token for fetching the next page of intent versions. If the response to this call is truncated, Amazon Lex returns a pagination token in the response. To fetch the next page of versions, specify the pagination token in the next request.

httpStatus :: Int

The response's http status code.

newGetIntentVersionsResponse

Arguments

:: Int
$sel:httpStatus:GetIntentVersionsResponse'

-> GetIntentVersionsResponse

Create a value of GetIntentVersionsResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:intents:GetIntentVersionsResponse', getIntentVersionsResponse_intents - An array of IntentMetadata objects, one for each numbered version of the intent plus one for the $LATEST version.

$sel:nextToken:GetIntentVersions', getIntentVersionsResponse_nextToken - A pagination token for fetching the next page of intent versions. If the response to this call is truncated, Amazon Lex returns a pagination token in the response. To fetch the next page of versions, specify the pagination token in the next request.

$sel:httpStatus:GetIntentVersionsResponse', getIntentVersionsResponse_httpStatus - The response's http status code.

Response Lenses

getIntentVersionsResponse_intents :: Lens' GetIntentVersionsResponse (Maybe [IntentMetadata])

An array of IntentMetadata objects, one for each numbered version of the intent plus one for the $LATEST version.

getIntentVersionsResponse_nextToken :: Lens' GetIntentVersionsResponse (Maybe Text)

A pagination token for fetching the next page of intent versions. If the response to this call is truncated, Amazon Lex returns a pagination token in the response. To fetch the next page of versions, specify the pagination token in the next request.

getIntentVersionsResponse_httpStatus :: Lens' GetIntentVersionsResponse Int

The response's http status code.