Amazonka.LexModels.GetMigration

Description

Provides details about an ongoing or complete migration from an Amazon Lex V1 bot to an Amazon Lex V2 bot. Use this operation to view the migration alerts and warnings related to the migration.

Creating a Request

data GetMigration

See: newGetMigration smart constructor.

Constructors

GetMigration'

Fields

migrationId :: Text

The unique identifier of the migration to view. The migrationID is returned by the operation.

newGetMigration

Arguments

:: Text
$sel:migrationId:GetMigration'

-> GetMigration

Create a value of GetMigration with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:migrationId:GetMigration', getMigration_migrationId - The unique identifier of the migration to view. The migrationID is returned by the operation.

Request Lenses

getMigration_migrationId :: Lens' GetMigration Text

The unique identifier of the migration to view. The migrationID is returned by the operation.

Destructuring the Response

data GetMigrationResponse

See: newGetMigrationResponse smart constructor.

Constructors

GetMigrationResponse'

Fields

v1BotVersion :: Maybe Text

The version of the Amazon Lex V1 bot migrated to Amazon Lex V2.

migrationStrategy :: Maybe MigrationStrategy

The strategy used to conduct the migration.

CREATE_NEW - Creates a new Amazon Lex V2 bot and migrates the Amazon Lex V1 bot to the new bot.

UPDATE_EXISTING - Overwrites the existing Amazon Lex V2 bot metadata and the locale being migrated. It doesn't change any other locales in the Amazon Lex V2 bot. If the locale doesn't exist, a new locale is created in the Amazon Lex V2 bot.

migrationTimestamp :: Maybe POSIX

The date and time that the migration started.

alerts :: Maybe [MigrationAlert]

A list of alerts and warnings that indicate issues with the migration for the Amazon Lex V1 bot to Amazon Lex V2. You receive a warning when an Amazon Lex V1 feature has a different implementation if Amazon Lex V2.

For more information, see Migrating a bot in the Amazon Lex V2 developer guide.

migrationStatus :: Maybe MigrationStatus

Indicates the status of the migration. When the status is COMPLETE the migration is finished and the bot is available in Amazon Lex V2. There may be alerts and warnings that need to be resolved to complete the migration.

v2BotId :: Maybe Text

The unique identifier of the Amazon Lex V2 bot that the Amazon Lex V1 is being migrated to.

v1BotLocale :: Maybe Locale

The locale of the Amazon Lex V1 bot migrated to Amazon Lex V2.

v1BotName :: Maybe Text

The name of the Amazon Lex V1 bot migrated to Amazon Lex V2.

v2BotRole :: Maybe Text

The IAM role that Amazon Lex uses to run the Amazon Lex V2 bot.

migrationId :: Maybe Text

The unique identifier of the migration. This is the same as the identifier used when calling the GetMigration operation.

httpStatus :: Int

The response's http status code.

newGetMigrationResponse

Arguments

:: Int
$sel:httpStatus:GetMigrationResponse'

-> GetMigrationResponse

Create a value of GetMigrationResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:v1BotVersion:GetMigrationResponse', getMigrationResponse_v1BotVersion - The version of the Amazon Lex V1 bot migrated to Amazon Lex V2.

$sel:migrationStrategy:GetMigrationResponse', getMigrationResponse_migrationStrategy - The strategy used to conduct the migration.

CREATE_NEW - Creates a new Amazon Lex V2 bot and migrates the Amazon Lex V1 bot to the new bot.

UPDATE_EXISTING - Overwrites the existing Amazon Lex V2 bot metadata and the locale being migrated. It doesn't change any other locales in the Amazon Lex V2 bot. If the locale doesn't exist, a new locale is created in the Amazon Lex V2 bot.

$sel:migrationTimestamp:GetMigrationResponse', getMigrationResponse_migrationTimestamp - The date and time that the migration started.

$sel:alerts:GetMigrationResponse', getMigrationResponse_alerts - A list of alerts and warnings that indicate issues with the migration for the Amazon Lex V1 bot to Amazon Lex V2. You receive a warning when an Amazon Lex V1 feature has a different implementation if Amazon Lex V2.

For more information, see Migrating a bot in the Amazon Lex V2 developer guide.

$sel:migrationStatus:GetMigrationResponse', getMigrationResponse_migrationStatus - Indicates the status of the migration. When the status is COMPLETE the migration is finished and the bot is available in Amazon Lex V2. There may be alerts and warnings that need to be resolved to complete the migration.

$sel:v2BotId:GetMigrationResponse', getMigrationResponse_v2BotId - The unique identifier of the Amazon Lex V2 bot that the Amazon Lex V1 is being migrated to.

$sel:v1BotLocale:GetMigrationResponse', getMigrationResponse_v1BotLocale - The locale of the Amazon Lex V1 bot migrated to Amazon Lex V2.

$sel:v1BotName:GetMigrationResponse', getMigrationResponse_v1BotName - The name of the Amazon Lex V1 bot migrated to Amazon Lex V2.

$sel:v2BotRole:GetMigrationResponse', getMigrationResponse_v2BotRole - The IAM role that Amazon Lex uses to run the Amazon Lex V2 bot.

$sel:migrationId:GetMigration', getMigrationResponse_migrationId - The unique identifier of the migration. This is the same as the identifier used when calling the GetMigration operation.

$sel:httpStatus:GetMigrationResponse', getMigrationResponse_httpStatus - The response's http status code.

Response Lenses

getMigrationResponse_v1BotVersion :: Lens' GetMigrationResponse (Maybe Text)

The version of the Amazon Lex V1 bot migrated to Amazon Lex V2.

getMigrationResponse_migrationStrategy :: Lens' GetMigrationResponse (Maybe MigrationStrategy)

The strategy used to conduct the migration.

CREATE_NEW - Creates a new Amazon Lex V2 bot and migrates the Amazon Lex V1 bot to the new bot.

UPDATE_EXISTING - Overwrites the existing Amazon Lex V2 bot metadata and the locale being migrated. It doesn't change any other locales in the Amazon Lex V2 bot. If the locale doesn't exist, a new locale is created in the Amazon Lex V2 bot.

getMigrationResponse_migrationTimestamp :: Lens' GetMigrationResponse (Maybe UTCTime)

The date and time that the migration started.

getMigrationResponse_alerts :: Lens' GetMigrationResponse (Maybe [MigrationAlert])

A list of alerts and warnings that indicate issues with the migration for the Amazon Lex V1 bot to Amazon Lex V2. You receive a warning when an Amazon Lex V1 feature has a different implementation if Amazon Lex V2.

For more information, see Migrating a bot in the Amazon Lex V2 developer guide.

getMigrationResponse_migrationStatus :: Lens' GetMigrationResponse (Maybe MigrationStatus)

Indicates the status of the migration. When the status is COMPLETE the migration is finished and the bot is available in Amazon Lex V2. There may be alerts and warnings that need to be resolved to complete the migration.

getMigrationResponse_v2BotId :: Lens' GetMigrationResponse (Maybe Text)

The unique identifier of the Amazon Lex V2 bot that the Amazon Lex V1 is being migrated to.

getMigrationResponse_v1BotLocale :: Lens' GetMigrationResponse (Maybe Locale)

The locale of the Amazon Lex V1 bot migrated to Amazon Lex V2.

getMigrationResponse_v1BotName :: Lens' GetMigrationResponse (Maybe Text)

The name of the Amazon Lex V1 bot migrated to Amazon Lex V2.

getMigrationResponse_v2BotRole :: Lens' GetMigrationResponse (Maybe Text)

The IAM role that Amazon Lex uses to run the Amazon Lex V2 bot.

getMigrationResponse_migrationId :: Lens' GetMigrationResponse (Maybe Text)

The unique identifier of the migration. This is the same as the identifier used when calling the GetMigration operation.

getMigrationResponse_httpStatus :: Lens' GetMigrationResponse Int

The response's http status code.