Amazonka.LexModels.GetUtterancesView

Description

Use the GetUtterancesView operation to get information about the utterances that your users have made to your bot. You can use this list to tune the utterances that your bot responds to.

For example, say that you have created a bot to order flowers. After your users have used your bot for a while, use the GetUtterancesView operation to see the requests that they have made and whether they have been successful. You might find that the utterance "I want flowers" is not being recognized. You could add this utterance to the OrderFlowers intent so that your bot recognizes that utterance.

After you publish a new version of a bot, you can get information about the old version and the new so that you can compare the performance across the two versions.

Utterance statistics are generated once a day. Data is available for the last 15 days. You can request information for up to 5 versions of your bot in each request. Amazon Lex returns the most frequent utterances received by the bot in the last 15 days. The response contains information about a maximum of 100 utterances for each version.

If you set childDirected field to true when you created your bot, if you are using slot obfuscation with one or more slots, or if you opted out of participating in improving Amazon Lex, utterances are not available.

This operation requires permissions for the lex:GetUtterancesView action.

Creating a Request

data GetUtterancesView

See: newGetUtterancesView smart constructor.

Constructors

GetUtterancesView'

Fields

botName :: Text

The name of the bot for which utterance information should be returned.

botVersions :: NonEmpty Text

An array of bot versions for which utterance information should be returned. The limit is 5 versions per request.

statusType :: StatusType

To return utterances that were recognized and handled, use Detected. To return utterances that were not recognized, use Missed.

newGetUtterancesView

Arguments

:: Text
$sel:botName:GetUtterancesView'

-> NonEmpty Text
$sel:botVersions:GetUtterancesView'

-> StatusType
$sel:statusType:GetUtterancesView'

-> GetUtterancesView

Create a value of GetUtterancesView with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:botName:GetUtterancesView', getUtterancesView_botName - The name of the bot for which utterance information should be returned.

$sel:botVersions:GetUtterancesView', getUtterancesView_botVersions - An array of bot versions for which utterance information should be returned. The limit is 5 versions per request.

$sel:statusType:GetUtterancesView', getUtterancesView_statusType - To return utterances that were recognized and handled, use Detected. To return utterances that were not recognized, use Missed.

Request Lenses

getUtterancesView_botName :: Lens' GetUtterancesView Text

The name of the bot for which utterance information should be returned.

getUtterancesView_botVersions :: Lens' GetUtterancesView (NonEmpty Text)

An array of bot versions for which utterance information should be returned. The limit is 5 versions per request.

getUtterancesView_statusType :: Lens' GetUtterancesView StatusType

To return utterances that were recognized and handled, use Detected. To return utterances that were not recognized, use Missed.

Destructuring the Response

data GetUtterancesViewResponse

See: newGetUtterancesViewResponse smart constructor.

Constructors

GetUtterancesViewResponse'

Fields

botName :: Maybe Text

The name of the bot for which utterance information was returned.

utterances :: Maybe [UtteranceList]

An array of UtteranceList objects, each containing a list of UtteranceData objects describing the utterances that were processed by your bot. The response contains a maximum of 100 UtteranceData objects for each version. Amazon Lex returns the most frequent utterances received by the bot in the last 15 days.

httpStatus :: Int

The response's http status code.

newGetUtterancesViewResponse

Arguments

:: Int
$sel:httpStatus:GetUtterancesViewResponse'

-> GetUtterancesViewResponse

Create a value of GetUtterancesViewResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:botName:GetUtterancesView', getUtterancesViewResponse_botName - The name of the bot for which utterance information was returned.

$sel:utterances:GetUtterancesViewResponse', getUtterancesViewResponse_utterances - An array of UtteranceList objects, each containing a list of UtteranceData objects describing the utterances that were processed by your bot. The response contains a maximum of 100 UtteranceData objects for each version. Amazon Lex returns the most frequent utterances received by the bot in the last 15 days.

$sel:httpStatus:GetUtterancesViewResponse', getUtterancesViewResponse_httpStatus - The response's http status code.

Response Lenses

getUtterancesViewResponse_botName :: Lens' GetUtterancesViewResponse (Maybe Text)

The name of the bot for which utterance information was returned.

getUtterancesViewResponse_utterances :: Lens' GetUtterancesViewResponse (Maybe [UtteranceList])

An array of UtteranceList objects, each containing a list of UtteranceData objects describing the utterances that were processed by your bot. The response contains a maximum of 100 UtteranceData objects for each version. Amazon Lex returns the most frequent utterances received by the bot in the last 15 days.

getUtterancesViewResponse_httpStatus :: Lens' GetUtterancesViewResponse Int

The response's http status code.