Amazonka.LexModels.StartMigration

Description

Starts migrating a bot from Amazon Lex V1 to Amazon Lex V2. Migrate your bot when you want to take advantage of the new features of Amazon Lex V2.

For more information, see Migrating a bot in the Amazon Lex developer guide.

Creating a Request

data StartMigration

See: newStartMigration smart constructor.

Constructors

StartMigration'

Fields

v1BotName :: Text

The name of the Amazon Lex V1 bot that you are migrating to Amazon Lex V2.

v1BotVersion :: Text

The version of the bot to migrate to Amazon Lex V2. You can migrate the $LATEST version as well as any numbered version.

v2BotName :: Text

The name of the Amazon Lex V2 bot that you are migrating the Amazon Lex V1 bot to.

If the Amazon Lex V2 bot doesn't exist, you must use the CREATE_NEW migration strategy.

If the Amazon Lex V2 bot exists, you must use the UPDATE_EXISTING migration strategy to change the contents of the Amazon Lex V2 bot.

v2BotRole :: Text

The IAM role that Amazon Lex uses to run the Amazon Lex V2 bot.

migrationStrategy :: MigrationStrategy

The strategy used to conduct the migration.

CREATE_NEW - Creates a new Amazon Lex V2 bot and migrates the Amazon Lex V1 bot to the new bot.

UPDATE_EXISTING - Overwrites the existing Amazon Lex V2 bot metadata and the locale being migrated. It doesn't change any other locales in the Amazon Lex V2 bot. If the locale doesn't exist, a new locale is created in the Amazon Lex V2 bot.

newStartMigration

Arguments

:: Text
$sel:v1BotName:StartMigration'

-> Text
$sel:v1BotVersion:StartMigration'

-> Text
$sel:v2BotName:StartMigration'

-> Text
$sel:v2BotRole:StartMigration'

-> MigrationStrategy
$sel:migrationStrategy:StartMigration'

-> StartMigration

Create a value of StartMigration with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:v1BotName:StartMigration', startMigration_v1BotName - The name of the Amazon Lex V1 bot that you are migrating to Amazon Lex V2.

$sel:v1BotVersion:StartMigration', startMigration_v1BotVersion - The version of the bot to migrate to Amazon Lex V2. You can migrate the $LATEST version as well as any numbered version.

$sel:v2BotName:StartMigration', startMigration_v2BotName - The name of the Amazon Lex V2 bot that you are migrating the Amazon Lex V1 bot to.

If the Amazon Lex V2 bot doesn't exist, you must use the CREATE_NEW migration strategy.

If the Amazon Lex V2 bot exists, you must use the UPDATE_EXISTING migration strategy to change the contents of the Amazon Lex V2 bot.

$sel:v2BotRole:StartMigration', startMigration_v2BotRole - The IAM role that Amazon Lex uses to run the Amazon Lex V2 bot.

$sel:migrationStrategy:StartMigration', startMigration_migrationStrategy - The strategy used to conduct the migration.

CREATE_NEW - Creates a new Amazon Lex V2 bot and migrates the Amazon Lex V1 bot to the new bot.

UPDATE_EXISTING - Overwrites the existing Amazon Lex V2 bot metadata and the locale being migrated. It doesn't change any other locales in the Amazon Lex V2 bot. If the locale doesn't exist, a new locale is created in the Amazon Lex V2 bot.

Request Lenses

startMigration_v1BotName :: Lens' StartMigration Text

The name of the Amazon Lex V1 bot that you are migrating to Amazon Lex V2.

startMigration_v1BotVersion :: Lens' StartMigration Text

The version of the bot to migrate to Amazon Lex V2. You can migrate the $LATEST version as well as any numbered version.

startMigration_v2BotName :: Lens' StartMigration Text

The name of the Amazon Lex V2 bot that you are migrating the Amazon Lex V1 bot to.

If the Amazon Lex V2 bot doesn't exist, you must use the CREATE_NEW migration strategy.

If the Amazon Lex V2 bot exists, you must use the UPDATE_EXISTING migration strategy to change the contents of the Amazon Lex V2 bot.

startMigration_v2BotRole :: Lens' StartMigration Text

The IAM role that Amazon Lex uses to run the Amazon Lex V2 bot.

startMigration_migrationStrategy :: Lens' StartMigration MigrationStrategy

The strategy used to conduct the migration.

CREATE_NEW - Creates a new Amazon Lex V2 bot and migrates the Amazon Lex V1 bot to the new bot.

UPDATE_EXISTING - Overwrites the existing Amazon Lex V2 bot metadata and the locale being migrated. It doesn't change any other locales in the Amazon Lex V2 bot. If the locale doesn't exist, a new locale is created in the Amazon Lex V2 bot.

Destructuring the Response

data StartMigrationResponse

See: newStartMigrationResponse smart constructor.

Constructors

StartMigrationResponse'

Fields

v1BotVersion :: Maybe Text

The version of the bot to migrate to Amazon Lex V2.

migrationStrategy :: Maybe MigrationStrategy

The strategy used to conduct the migration.

migrationTimestamp :: Maybe POSIX

The date and time that the migration started.

v2BotId :: Maybe Text

The unique identifier for the Amazon Lex V2 bot.

v1BotLocale :: Maybe Locale

The locale used for the Amazon Lex V1 bot.

v1BotName :: Maybe Text

The name of the Amazon Lex V1 bot that you are migrating to Amazon Lex V2.

v2BotRole :: Maybe Text

The IAM role that Amazon Lex uses to run the Amazon Lex V2 bot.

migrationId :: Maybe Text

The unique identifier that Amazon Lex assigned to the migration.

httpStatus :: Int

The response's http status code.

newStartMigrationResponse

Arguments

:: Int
$sel:httpStatus:StartMigrationResponse'

-> StartMigrationResponse

Create a value of StartMigrationResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:v1BotVersion:StartMigration', startMigrationResponse_v1BotVersion - The version of the bot to migrate to Amazon Lex V2.

$sel:migrationStrategy:StartMigration', startMigrationResponse_migrationStrategy - The strategy used to conduct the migration.

$sel:migrationTimestamp:StartMigrationResponse', startMigrationResponse_migrationTimestamp - The date and time that the migration started.

$sel:v2BotId:StartMigrationResponse', startMigrationResponse_v2BotId - The unique identifier for the Amazon Lex V2 bot.

$sel:v1BotLocale:StartMigrationResponse', startMigrationResponse_v1BotLocale - The locale used for the Amazon Lex V1 bot.

$sel:v1BotName:StartMigration', startMigrationResponse_v1BotName - The name of the Amazon Lex V1 bot that you are migrating to Amazon Lex V2.

$sel:v2BotRole:StartMigration', startMigrationResponse_v2BotRole - The IAM role that Amazon Lex uses to run the Amazon Lex V2 bot.

$sel:migrationId:StartMigrationResponse', startMigrationResponse_migrationId - The unique identifier that Amazon Lex assigned to the migration.

$sel:httpStatus:StartMigrationResponse', startMigrationResponse_httpStatus - The response's http status code.

Response Lenses

startMigrationResponse_v1BotVersion :: Lens' StartMigrationResponse (Maybe Text)

The version of the bot to migrate to Amazon Lex V2.

startMigrationResponse_migrationStrategy :: Lens' StartMigrationResponse (Maybe MigrationStrategy)

The strategy used to conduct the migration.

startMigrationResponse_migrationTimestamp :: Lens' StartMigrationResponse (Maybe UTCTime)

The date and time that the migration started.

startMigrationResponse_v2BotId :: Lens' StartMigrationResponse (Maybe Text)

The unique identifier for the Amazon Lex V2 bot.

startMigrationResponse_v1BotLocale :: Lens' StartMigrationResponse (Maybe Locale)

The locale used for the Amazon Lex V1 bot.

startMigrationResponse_v1BotName :: Lens' StartMigrationResponse (Maybe Text)

The name of the Amazon Lex V1 bot that you are migrating to Amazon Lex V2.

startMigrationResponse_v2BotRole :: Lens' StartMigrationResponse (Maybe Text)

The IAM role that Amazon Lex uses to run the Amazon Lex V2 bot.

startMigrationResponse_migrationId :: Lens' StartMigrationResponse (Maybe Text)

The unique identifier that Amazon Lex assigned to the migration.

startMigrationResponse_httpStatus :: Lens' StartMigrationResponse Int

The response's http status code.