Amazonka.LexModels.Types.BotMetadata

Description

Documentation

data BotMetadata

Provides information about a bot. .

See: newBotMetadata smart constructor.

Constructors

BotMetadata'

Fields

status :: Maybe LexStatus

The status of the bot.

createdDate :: Maybe POSIX

The date that the bot was created.

name :: Maybe Text

The name of the bot.

version :: Maybe Text

The version of the bot. For a new bot, the version is always $LATEST.

lastUpdatedDate :: Maybe POSIX

The date that the bot was updated. When you create a bot, the creation date and last updated date are the same.

description :: Maybe Text

A description of the bot.

newBotMetadata :: BotMetadata

Create a value of BotMetadata with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:status:BotMetadata', botMetadata_status - The status of the bot.

$sel:createdDate:BotMetadata', botMetadata_createdDate - The date that the bot was created.

$sel:name:BotMetadata', botMetadata_name - The name of the bot.

$sel:version:BotMetadata', botMetadata_version - The version of the bot. For a new bot, the version is always $LATEST.

$sel:lastUpdatedDate:BotMetadata', botMetadata_lastUpdatedDate - The date that the bot was updated. When you create a bot, the creation date and last updated date are the same.

$sel:description:BotMetadata', botMetadata_description - A description of the bot.

botMetadata_status :: Lens' BotMetadata (Maybe LexStatus)

The status of the bot.

botMetadata_createdDate :: Lens' BotMetadata (Maybe UTCTime)

The date that the bot was created.

botMetadata_name :: Lens' BotMetadata (Maybe Text)

The name of the bot.

botMetadata_version :: Lens' BotMetadata (Maybe Text)

The version of the bot. For a new bot, the version is always $LATEST.

botMetadata_lastUpdatedDate :: Lens' BotMetadata (Maybe UTCTime)

The date that the bot was updated. When you create a bot, the creation date and last updated date are the same.

botMetadata_description :: Lens' BotMetadata (Maybe Text)

A description of the bot.