Amazonka.LexModels.Types.IntentMetadata

Description

Documentation

data IntentMetadata

Provides information about an intent.

See: newIntentMetadata smart constructor.

Constructors

IntentMetadata'

Fields

createdDate :: Maybe POSIX

The date that the intent was created.

name :: Maybe Text

The name of the intent.

version :: Maybe Text

The version of the intent.

lastUpdatedDate :: Maybe POSIX

The date that the intent was updated. When you create an intent, the creation date and last updated date are the same.

description :: Maybe Text

A description of the intent.

newIntentMetadata :: IntentMetadata

Create a value of IntentMetadata with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:createdDate:IntentMetadata', intentMetadata_createdDate - The date that the intent was created.

$sel:name:IntentMetadata', intentMetadata_name - The name of the intent.

$sel:version:IntentMetadata', intentMetadata_version - The version of the intent.

$sel:lastUpdatedDate:IntentMetadata', intentMetadata_lastUpdatedDate - The date that the intent was updated. When you create an intent, the creation date and last updated date are the same.

$sel:description:IntentMetadata', intentMetadata_description - A description of the intent.

intentMetadata_createdDate :: Lens' IntentMetadata (Maybe UTCTime)

The date that the intent was created.

intentMetadata_name :: Lens' IntentMetadata (Maybe Text)

The name of the intent.

intentMetadata_version :: Lens' IntentMetadata (Maybe Text)

The version of the intent.

intentMetadata_lastUpdatedDate :: Lens' IntentMetadata (Maybe UTCTime)

The date that the intent was updated. When you create an intent, the creation date and last updated date are the same.

intentMetadata_description :: Lens' IntentMetadata (Maybe Text)

A description of the intent.