Amazonka.LexModels.Types.Prompt

Description

Documentation

data Prompt

Obtains information from the user. To define a prompt, provide one or more messages and specify the number of attempts to get information from the user. If you provide more than one message, Amazon Lex chooses one of the messages to use to prompt the user. For more information, see how-it-works.

See: newPrompt smart constructor.

Constructors

Prompt'

Fields

responseCard :: Maybe Text

A response card. Amazon Lex uses this prompt at runtime, in the PostText API response. It substitutes session attributes and slot values for placeholders in the response card. For more information, see ex-resp-card.

messages :: NonEmpty Message

An array of objects, each of which provides a message string and its type. You can specify the message string in plain text or in Speech Synthesis Markup Language (SSML).

maxAttempts :: Natural

The number of times to prompt the user for information.

newPrompt

Arguments

:: NonEmpty Message
$sel:messages:Prompt'

-> Natural
$sel:maxAttempts:Prompt'

-> Prompt

Create a value of Prompt with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:responseCard:Prompt', prompt_responseCard - A response card. Amazon Lex uses this prompt at runtime, in the PostText API response. It substitutes session attributes and slot values for placeholders in the response card. For more information, see ex-resp-card.

$sel:messages:Prompt', prompt_messages - An array of objects, each of which provides a message string and its type. You can specify the message string in plain text or in Speech Synthesis Markup Language (SSML).

$sel:maxAttempts:Prompt', prompt_maxAttempts - The number of times to prompt the user for information.

prompt_responseCard :: Lens' Prompt (Maybe Text)

A response card. Amazon Lex uses this prompt at runtime, in the PostText API response. It substitutes session attributes and slot values for placeholders in the response card. For more information, see ex-resp-card.

prompt_messages :: Lens' Prompt (NonEmpty Message)

An array of objects, each of which provides a message string and its type. You can specify the message string in plain text or in Speech Synthesis Markup Language (SSML).

prompt_maxAttempts :: Lens' Prompt Natural

The number of times to prompt the user for information.