Amazonka.LexModels.Types.SlotTypeMetadata

Description

Documentation

data SlotTypeMetadata

Provides information about a slot type..

See: newSlotTypeMetadata smart constructor.

Constructors

SlotTypeMetadata'

Fields

createdDate :: Maybe POSIX

The date that the slot type was created.

name :: Maybe Text

The name of the slot type.

version :: Maybe Text

The version of the slot type.

lastUpdatedDate :: Maybe POSIX

The date that the slot type was updated. When you create a resource, the creation date and last updated date are the same.

description :: Maybe Text

A description of the slot type.

newSlotTypeMetadata :: SlotTypeMetadata

Create a value of SlotTypeMetadata with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:createdDate:SlotTypeMetadata', slotTypeMetadata_createdDate - The date that the slot type was created.

$sel:name:SlotTypeMetadata', slotTypeMetadata_name - The name of the slot type.

$sel:version:SlotTypeMetadata', slotTypeMetadata_version - The version of the slot type.

$sel:lastUpdatedDate:SlotTypeMetadata', slotTypeMetadata_lastUpdatedDate - The date that the slot type was updated. When you create a resource, the creation date and last updated date are the same.

$sel:description:SlotTypeMetadata', slotTypeMetadata_description - A description of the slot type.

slotTypeMetadata_createdDate :: Lens' SlotTypeMetadata (Maybe UTCTime)

The date that the slot type was created.

slotTypeMetadata_name :: Lens' SlotTypeMetadata (Maybe Text)

The name of the slot type.

slotTypeMetadata_version :: Lens' SlotTypeMetadata (Maybe Text)

The version of the slot type.

slotTypeMetadata_lastUpdatedDate :: Lens' SlotTypeMetadata (Maybe UTCTime)

The date that the slot type was updated. When you create a resource, the creation date and last updated date are the same.

slotTypeMetadata_description :: Lens' SlotTypeMetadata (Maybe Text)

A description of the slot type.