Amazonka.LexRuntime.GetSession

Description

Returns session information for a specified bot, alias, and user ID.

Creating a Request

data GetSession

See: newGetSession smart constructor.

Constructors

GetSession'

Fields

checkpointLabelFilter :: Maybe Text

A string used to filter the intents returned in the recentIntentSummaryView structure.

When you specify a filter, only intents with their checkpointLabel field set to that string are returned.

botName :: Text

The name of the bot that contains the session data.

botAlias :: Text

The alias in use for the bot that contains the session data.

userId :: Text

The ID of the client application user. Amazon Lex uses this to identify a user's conversation with your bot.

newGetSession

Arguments

:: Text
$sel:botName:GetSession'

-> Text
$sel:botAlias:GetSession'

-> Text
$sel:userId:GetSession'

-> GetSession

Create a value of GetSession with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:checkpointLabelFilter:GetSession', getSession_checkpointLabelFilter - A string used to filter the intents returned in the recentIntentSummaryView structure.

When you specify a filter, only intents with their checkpointLabel field set to that string are returned.

$sel:botName:GetSession', getSession_botName - The name of the bot that contains the session data.

$sel:botAlias:GetSession', getSession_botAlias - The alias in use for the bot that contains the session data.

$sel:userId:GetSession', getSession_userId - The ID of the client application user. Amazon Lex uses this to identify a user's conversation with your bot.

Request Lenses

getSession_checkpointLabelFilter :: Lens' GetSession (Maybe Text)

A string used to filter the intents returned in the recentIntentSummaryView structure.

When you specify a filter, only intents with their checkpointLabel field set to that string are returned.

getSession_botName :: Lens' GetSession Text

The name of the bot that contains the session data.

getSession_botAlias :: Lens' GetSession Text

The alias in use for the bot that contains the session data.

getSession_userId :: Lens' GetSession Text

The ID of the client application user. Amazon Lex uses this to identify a user's conversation with your bot.

Destructuring the Response

data GetSessionResponse

See: newGetSessionResponse smart constructor.

Constructors

GetSessionResponse'

Fields

activeContexts :: Maybe (Sensitive [ActiveContext])

A list of active contexts for the session. A context can be set when an intent is fulfilled or by calling the PostContent, PostText, or PutSession operation.

You can use a context to control the intents that can follow up an intent, or to modify the operation of your application.

sessionId :: Maybe Text

A unique identifier for the session.

recentIntentSummaryView :: Maybe [IntentSummary]

An array of information about the intents used in the session. The array can contain a maximum of three summaries. If more than three intents are used in the session, the recentIntentSummaryView operation contains information about the last three intents used.

If you set the checkpointLabelFilter parameter in the request, the array contains only the intents with the specified label.

dialogAction :: Maybe DialogAction

Describes the current state of the bot.

sessionAttributes :: Maybe (Sensitive (HashMap Text Text))

Map of key/value pairs representing the session-specific context information. It contains application information passed between Amazon Lex and a client application.

httpStatus :: Int

The response's http status code.

newGetSessionResponse

Arguments

:: Int
$sel:httpStatus:GetSessionResponse'

-> GetSessionResponse

Create a value of GetSessionResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:activeContexts:GetSessionResponse', getSessionResponse_activeContexts - A list of active contexts for the session. A context can be set when an intent is fulfilled or by calling the PostContent, PostText, or PutSession operation.

You can use a context to control the intents that can follow up an intent, or to modify the operation of your application.

$sel:sessionId:GetSessionResponse', getSessionResponse_sessionId - A unique identifier for the session.

$sel:recentIntentSummaryView:GetSessionResponse', getSessionResponse_recentIntentSummaryView - An array of information about the intents used in the session. The array can contain a maximum of three summaries. If more than three intents are used in the session, the recentIntentSummaryView operation contains information about the last three intents used.

If you set the checkpointLabelFilter parameter in the request, the array contains only the intents with the specified label.

$sel:dialogAction:GetSessionResponse', getSessionResponse_dialogAction - Describes the current state of the bot.

$sel:sessionAttributes:GetSessionResponse', getSessionResponse_sessionAttributes - Map of key/value pairs representing the session-specific context information. It contains application information passed between Amazon Lex and a client application.

$sel:httpStatus:GetSessionResponse', getSessionResponse_httpStatus - The response's http status code.

Response Lenses

getSessionResponse_activeContexts :: Lens' GetSessionResponse (Maybe [ActiveContext])

A list of active contexts for the session. A context can be set when an intent is fulfilled or by calling the PostContent, PostText, or PutSession operation.

You can use a context to control the intents that can follow up an intent, or to modify the operation of your application.

getSessionResponse_sessionId :: Lens' GetSessionResponse (Maybe Text)

A unique identifier for the session.

getSessionResponse_recentIntentSummaryView :: Lens' GetSessionResponse (Maybe [IntentSummary])

An array of information about the intents used in the session. The array can contain a maximum of three summaries. If more than three intents are used in the session, the recentIntentSummaryView operation contains information about the last three intents used.

If you set the checkpointLabelFilter parameter in the request, the array contains only the intents with the specified label.

getSessionResponse_dialogAction :: Lens' GetSessionResponse (Maybe DialogAction)

Describes the current state of the bot.

getSessionResponse_sessionAttributes :: Lens' GetSessionResponse (Maybe (HashMap Text Text))

Map of key/value pairs representing the session-specific context information. It contains application information passed between Amazon Lex and a client application.

getSessionResponse_httpStatus :: Lens' GetSessionResponse Int

The response's http status code.