Amazonka.LexRuntime.Types

Description

Service Configuration

defaultService :: Service

API version 2016-11-28 of the Amazon Lex Runtime Service SDK configuration.

Errors

_NotAcceptableException :: AsError a => Getting (First ServiceError) a ServiceError

The accept header in the request does not have a valid value.

_DependencyFailedException :: AsError a => Getting (First ServiceError) a ServiceError

One of the dependencies, such as AWS Lambda or Amazon Polly, threw an exception. For example,

If Amazon Lex does not have sufficient permissions to call a Lambda function.

If a Lambda function takes longer than 30 seconds to execute.

If a fulfillment Lambda function returns a Delegate dialog action without removing any slot values.

_UnsupportedMediaTypeException :: AsError a => Getting (First ServiceError) a ServiceError

The Content-Type header (PostContent API) has an invalid value.

_ConflictException :: AsError a => Getting (First ServiceError) a ServiceError

Two clients are using the same AWS account, Amazon Lex bot, and user ID.

_NotFoundException :: AsError a => Getting (First ServiceError) a ServiceError

The resource (such as the Amazon Lex bot or an alias) that is referred to is not found.

_RequestTimeoutException :: AsError a => Getting (First ServiceError) a ServiceError

The input speech is too long.

_LoopDetectedException :: AsError a => Getting (First ServiceError) a ServiceError

This exception is not used.

_InternalFailureException :: AsError a => Getting (First ServiceError) a ServiceError

Internal service error. Retry the call.

_BadGatewayException :: AsError a => Getting (First ServiceError) a ServiceError

Either the Amazon Lex bot is still building, or one of the dependent services (Amazon Polly, AWS Lambda) failed with an internal service error.

_BadRequestException :: AsError a => Getting (First ServiceError) a ServiceError

Request validation failed, there is no usable message in the context, or the bot build failed, is still in progress, or contains unbuilt changes.

_LimitExceededException :: AsError a => Getting (First ServiceError) a ServiceError

Exceeded a limit.

ConfirmationStatus

newtype ConfirmationStatus

Constructors

ConfirmationStatus'

Fields

fromConfirmationStatus :: Text

Bundled Patterns

pattern ConfirmationStatus_Confirmed :: ConfirmationStatus

pattern ConfirmationStatus_Denied :: ConfirmationStatus

pattern ConfirmationStatus_None :: ConfirmationStatus

ContentType

newtype ContentType

Constructors

ContentType'

Fields

fromContentType :: Text

Bundled Patterns

pattern ContentType_Application_vnd_amazonaws_card_generic :: ContentType

DialogActionType

newtype DialogActionType

Constructors

DialogActionType'

Fields

fromDialogActionType :: Text

Bundled Patterns

pattern DialogActionType_Close :: DialogActionType

pattern DialogActionType_ConfirmIntent :: DialogActionType

pattern DialogActionType_Delegate :: DialogActionType

pattern DialogActionType_ElicitIntent :: DialogActionType

pattern DialogActionType_ElicitSlot :: DialogActionType

DialogState

newtype DialogState

Constructors

DialogState'

Fields

fromDialogState :: Text

Bundled Patterns

pattern DialogState_ConfirmIntent :: DialogState

pattern DialogState_ElicitIntent :: DialogState

pattern DialogState_ElicitSlot :: DialogState

pattern DialogState_Failed :: DialogState

pattern DialogState_Fulfilled :: DialogState

pattern DialogState_ReadyForFulfillment :: DialogState

FulfillmentState

newtype FulfillmentState

Constructors

FulfillmentState'

Fields

fromFulfillmentState :: Text

Bundled Patterns

pattern FulfillmentState_Failed :: FulfillmentState

pattern FulfillmentState_Fulfilled :: FulfillmentState

pattern FulfillmentState_ReadyForFulfillment :: FulfillmentState

MessageFormatType

newtype MessageFormatType

Constructors

MessageFormatType'

Fields

fromMessageFormatType :: Text

Bundled Patterns

pattern MessageFormatType_Composite :: MessageFormatType

pattern MessageFormatType_CustomPayload :: MessageFormatType

pattern MessageFormatType_PlainText :: MessageFormatType

pattern MessageFormatType_SSML :: MessageFormatType

ActiveContext

data ActiveContext

A context is a variable that contains information about the current state of the conversation between a user and Amazon Lex. Context can be set automatically by Amazon Lex when an intent is fulfilled, or it can be set at runtime using the PutContent, PutText, or PutSession operation.

See: newActiveContext smart constructor.

Constructors

ActiveContext'

Fields

name :: Text

The name of the context.

timeToLive :: ActiveContextTimeToLive

The length of time or number of turns that a context remains active.

parameters :: HashMap Text (Sensitive Text)

State variables for the current context. You can use these values as default values for slots in subsequent events.

newActiveContext

Arguments

:: Text
$sel:name:ActiveContext'

-> ActiveContextTimeToLive
$sel:timeToLive:ActiveContext'

-> ActiveContext

Create a value of ActiveContext with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:name:ActiveContext', activeContext_name - The name of the context.

$sel:timeToLive:ActiveContext', activeContext_timeToLive - The length of time or number of turns that a context remains active.

$sel:parameters:ActiveContext', activeContext_parameters - State variables for the current context. You can use these values as default values for slots in subsequent events.

activeContext_name :: Lens' ActiveContext Text

The name of the context.

activeContext_timeToLive :: Lens' ActiveContext ActiveContextTimeToLive

The length of time or number of turns that a context remains active.

activeContext_parameters :: Lens' ActiveContext (HashMap Text Text)

State variables for the current context. You can use these values as default values for slots in subsequent events.

ActiveContextTimeToLive

data ActiveContextTimeToLive

The length of time or number of turns that a context remains active.

See: newActiveContextTimeToLive smart constructor.

Constructors

ActiveContextTimeToLive'

Fields

turnsToLive :: Maybe Natural

The number of conversation turns that the context should be active. A conversation turn is one PostContent or PostText request and the corresponding response from Amazon Lex.

timeToLiveInSeconds :: Maybe Natural

The number of seconds that the context should be active after it is first sent in a PostContent or PostText response. You can set the value between 5 and 86,400 seconds (24 hours).

newActiveContextTimeToLive :: ActiveContextTimeToLive

Create a value of ActiveContextTimeToLive with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:turnsToLive:ActiveContextTimeToLive', activeContextTimeToLive_turnsToLive - The number of conversation turns that the context should be active. A conversation turn is one PostContent or PostText request and the corresponding response from Amazon Lex.

$sel:timeToLiveInSeconds:ActiveContextTimeToLive', activeContextTimeToLive_timeToLiveInSeconds - The number of seconds that the context should be active after it is first sent in a PostContent or PostText response. You can set the value between 5 and 86,400 seconds (24 hours).

activeContextTimeToLive_turnsToLive :: Lens' ActiveContextTimeToLive (Maybe Natural)

The number of conversation turns that the context should be active. A conversation turn is one PostContent or PostText request and the corresponding response from Amazon Lex.

activeContextTimeToLive_timeToLiveInSeconds :: Lens' ActiveContextTimeToLive (Maybe Natural)

The number of seconds that the context should be active after it is first sent in a PostContent or PostText response. You can set the value between 5 and 86,400 seconds (24 hours).

Button

data Button

Represents an option to be shown on the client platform (Facebook, Slack, etc.)

See: newButton smart constructor.

Constructors

Button'

Fields

text :: Text

Text that is visible to the user on the button.

value :: Text

The value sent to Amazon Lex when a user chooses the button. For example, consider button text "NYC." When the user chooses the button, the value sent can be "New York City."

newButton

Arguments

:: Text
$sel:text:Button'

-> Text
$sel:value:Button'

-> Button

Create a value of Button with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:text:Button', button_text - Text that is visible to the user on the button.

$sel:value:Button', button_value - The value sent to Amazon Lex when a user chooses the button. For example, consider button text "NYC." When the user chooses the button, the value sent can be "New York City."

button_text :: Lens' Button Text

Text that is visible to the user on the button.

button_value :: Lens' Button Text

The value sent to Amazon Lex when a user chooses the button. For example, consider button text "NYC." When the user chooses the button, the value sent can be "New York City."

DialogAction

data DialogAction

Describes the next action that the bot should take in its interaction with the user and provides information about the context in which the action takes place. Use the DialogAction data type to set the interaction to a specific state, or to return the interaction to a previous state.

See: newDialogAction smart constructor.

Constructors

DialogAction'

Fields

slots :: Maybe (Sensitive (HashMap Text Text))

Map of the slots that have been gathered and their values.

intentName :: Maybe Text

The name of the intent.

fulfillmentState :: Maybe FulfillmentState

The fulfillment state of the intent. The possible values are:

Failed - The Lambda function associated with the intent failed to fulfill the intent.

Fulfilled - The intent has fulfilled by the Lambda function associated with the intent.

ReadyForFulfillment - All of the information necessary for the intent is present and the intent ready to be fulfilled by the client application.

messageFormat :: Maybe MessageFormatType

PlainText - The message contains plain UTF-8 text.

CustomPayload - The message is a custom format for the client.

SSML - The message contains text formatted for voice output.

Composite - The message contains an escaped JSON object containing one or more messages. For more information, see Message Groups.

message :: Maybe (Sensitive Text)

The message that should be shown to the user. If you don't specify a message, Amazon Lex will use the message configured for the intent.

slotToElicit :: Maybe Text

The name of the slot that should be elicited from the user.

type' :: DialogActionType

The next action that the bot should take in its interaction with the user. The possible values are:

ConfirmIntent - The next action is asking the user if the intent is complete and ready to be fulfilled. This is a yes/no question such as "Place the order?"

Close - Indicates that the there will not be a response from the user. For example, the statement "Your order has been placed" does not require a response.

Delegate - The next action is determined by Amazon Lex.

ElicitIntent - The next action is to determine the intent that the user wants to fulfill.

ElicitSlot - The next action is to elicit a slot value from the user.

newDialogAction

Arguments

:: DialogActionType
$sel:type':DialogAction'

-> DialogAction

Create a value of DialogAction with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:slots:DialogAction', dialogAction_slots - Map of the slots that have been gathered and their values.

$sel:intentName:DialogAction', dialogAction_intentName - The name of the intent.

$sel:fulfillmentState:DialogAction', dialogAction_fulfillmentState - The fulfillment state of the intent. The possible values are:

Failed - The Lambda function associated with the intent failed to fulfill the intent.

Fulfilled - The intent has fulfilled by the Lambda function associated with the intent.

ReadyForFulfillment - All of the information necessary for the intent is present and the intent ready to be fulfilled by the client application.

$sel:messageFormat:DialogAction', dialogAction_messageFormat - - PlainText - The message contains plain UTF-8 text.

CustomPayload - The message is a custom format for the client.

SSML - The message contains text formatted for voice output.

Composite - The message contains an escaped JSON object containing one or more messages. For more information, see Message Groups.

$sel:message:DialogAction', dialogAction_message - The message that should be shown to the user. If you don't specify a message, Amazon Lex will use the message configured for the intent.

$sel:slotToElicit:DialogAction', dialogAction_slotToElicit - The name of the slot that should be elicited from the user.

$sel:type':DialogAction', dialogAction_type - The next action that the bot should take in its interaction with the user. The possible values are:

ConfirmIntent - The next action is asking the user if the intent is complete and ready to be fulfilled. This is a yes/no question such as "Place the order?"

Close - Indicates that the there will not be a response from the user. For example, the statement "Your order has been placed" does not require a response.

Delegate - The next action is determined by Amazon Lex.

ElicitIntent - The next action is to determine the intent that the user wants to fulfill.

ElicitSlot - The next action is to elicit a slot value from the user.

dialogAction_slots :: Lens' DialogAction (Maybe (HashMap Text Text))

Map of the slots that have been gathered and their values.

dialogAction_intentName :: Lens' DialogAction (Maybe Text)

The name of the intent.

dialogAction_fulfillmentState :: Lens' DialogAction (Maybe FulfillmentState)

The fulfillment state of the intent. The possible values are:

Failed - The Lambda function associated with the intent failed to fulfill the intent.

Fulfilled - The intent has fulfilled by the Lambda function associated with the intent.

ReadyForFulfillment - All of the information necessary for the intent is present and the intent ready to be fulfilled by the client application.

dialogAction_messageFormat :: Lens' DialogAction (Maybe MessageFormatType)

PlainText - The message contains plain UTF-8 text.

CustomPayload - The message is a custom format for the client.

SSML - The message contains text formatted for voice output.

Composite - The message contains an escaped JSON object containing one or more messages. For more information, see Message Groups.

dialogAction_message :: Lens' DialogAction (Maybe Text)

The message that should be shown to the user. If you don't specify a message, Amazon Lex will use the message configured for the intent.

dialogAction_slotToElicit :: Lens' DialogAction (Maybe Text)

The name of the slot that should be elicited from the user.

dialogAction_type :: Lens' DialogAction DialogActionType

The next action that the bot should take in its interaction with the user. The possible values are:

ConfirmIntent - The next action is asking the user if the intent is complete and ready to be fulfilled. This is a yes/no question such as "Place the order?"

Close - Indicates that the there will not be a response from the user. For example, the statement "Your order has been placed" does not require a response.

Delegate - The next action is determined by Amazon Lex.

ElicitIntent - The next action is to determine the intent that the user wants to fulfill.

ElicitSlot - The next action is to elicit a slot value from the user.

GenericAttachment

data GenericAttachment

Represents an option rendered to the user when a prompt is shown. It could be an image, a button, a link, or text.

See: newGenericAttachment smart constructor.

Constructors

GenericAttachment'

Fields

buttons :: Maybe [Button]

The list of options to show to the user.

subTitle :: Maybe Text

The subtitle shown below the title.

imageUrl :: Maybe Text

The URL of an image that is displayed to the user.

attachmentLinkUrl :: Maybe Text

The URL of an attachment to the response card.

title :: Maybe Text

The title of the option.

newGenericAttachment :: GenericAttachment

Create a value of GenericAttachment with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:buttons:GenericAttachment', genericAttachment_buttons - The list of options to show to the user.

$sel:subTitle:GenericAttachment', genericAttachment_subTitle - The subtitle shown below the title.

$sel:imageUrl:GenericAttachment', genericAttachment_imageUrl - The URL of an image that is displayed to the user.

$sel:attachmentLinkUrl:GenericAttachment', genericAttachment_attachmentLinkUrl - The URL of an attachment to the response card.

$sel:title:GenericAttachment', genericAttachment_title - The title of the option.

genericAttachment_buttons :: Lens' GenericAttachment (Maybe [Button])

The list of options to show to the user.

genericAttachment_subTitle :: Lens' GenericAttachment (Maybe Text)

The subtitle shown below the title.

genericAttachment_imageUrl :: Lens' GenericAttachment (Maybe Text)

The URL of an image that is displayed to the user.

genericAttachment_attachmentLinkUrl :: Lens' GenericAttachment (Maybe Text)

The URL of an attachment to the response card.

genericAttachment_title :: Lens' GenericAttachment (Maybe Text)

The title of the option.

IntentConfidence

data IntentConfidence

Provides a score that indicates the confidence that Amazon Lex has that an intent is the one that satisfies the user's intent.

See: newIntentConfidence smart constructor.

Constructors

IntentConfidence'

Fields

score :: Maybe Double

A score that indicates how confident Amazon Lex is that an intent satisfies the user's intent. Ranges between 0.00 and 1.00. Higher scores indicate higher confidence.

newIntentConfidence :: IntentConfidence

Create a value of IntentConfidence with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:score:IntentConfidence', intentConfidence_score - A score that indicates how confident Amazon Lex is that an intent satisfies the user's intent. Ranges between 0.00 and 1.00. Higher scores indicate higher confidence.

intentConfidence_score :: Lens' IntentConfidence (Maybe Double)

A score that indicates how confident Amazon Lex is that an intent satisfies the user's intent. Ranges between 0.00 and 1.00. Higher scores indicate higher confidence.

IntentSummary

data IntentSummary

Provides information about the state of an intent. You can use this information to get the current state of an intent so that you can process the intent, or so that you can return the intent to its previous state.

See: newIntentSummary smart constructor.

Constructors

IntentSummary'

Fields

checkpointLabel :: Maybe Text

A user-defined label that identifies a particular intent. You can use this label to return to a previous intent.

Use the checkpointLabelFilter parameter of the GetSessionRequest operation to filter the intents returned by the operation to those with only the specified label.

slots :: Maybe (Sensitive (HashMap Text Text))

Map of the slots that have been gathered and their values.

intentName :: Maybe Text

The name of the intent.

fulfillmentState :: Maybe FulfillmentState

The fulfillment state of the intent. The possible values are:

Failed - The Lambda function associated with the intent failed to fulfill the intent.

Fulfilled - The intent has fulfilled by the Lambda function associated with the intent.

ReadyForFulfillment - All of the information necessary for the intent is present and the intent ready to be fulfilled by the client application.

confirmationStatus :: Maybe ConfirmationStatus

The status of the intent after the user responds to the confirmation prompt. If the user confirms the intent, Amazon Lex sets this field to Confirmed. If the user denies the intent, Amazon Lex sets this value to Denied. The possible values are:

Confirmed - The user has responded "Yes" to the confirmation prompt, confirming that the intent is complete and that it is ready to be fulfilled.

Denied - The user has responded "No" to the confirmation prompt.

None - The user has never been prompted for confirmation; or, the user was prompted but did not confirm or deny the prompt.

slotToElicit :: Maybe Text

The next slot to elicit from the user. If there is not slot to elicit, the field is blank.

dialogActionType :: DialogActionType

The next action that the bot should take in its interaction with the user. The possible values are:

ConfirmIntent - The next action is asking the user if the intent is complete and ready to be fulfilled. This is a yes/no question such as "Place the order?"

Close - Indicates that the there will not be a response from the user. For example, the statement "Your order has been placed" does not require a response.

ElicitIntent - The next action is to determine the intent that the user wants to fulfill.

ElicitSlot - The next action is to elicit a slot value from the user.

newIntentSummary

Arguments

:: DialogActionType
$sel:dialogActionType:IntentSummary'

-> IntentSummary

Create a value of IntentSummary with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:checkpointLabel:IntentSummary', intentSummary_checkpointLabel - A user-defined label that identifies a particular intent. You can use this label to return to a previous intent.

Use the checkpointLabelFilter parameter of the GetSessionRequest operation to filter the intents returned by the operation to those with only the specified label.

$sel:slots:IntentSummary', intentSummary_slots - Map of the slots that have been gathered and their values.

$sel:intentName:IntentSummary', intentSummary_intentName - The name of the intent.

$sel:fulfillmentState:IntentSummary', intentSummary_fulfillmentState - The fulfillment state of the intent. The possible values are:

Failed - The Lambda function associated with the intent failed to fulfill the intent.

Fulfilled - The intent has fulfilled by the Lambda function associated with the intent.

ReadyForFulfillment - All of the information necessary for the intent is present and the intent ready to be fulfilled by the client application.

$sel:confirmationStatus:IntentSummary', intentSummary_confirmationStatus - The status of the intent after the user responds to the confirmation prompt. If the user confirms the intent, Amazon Lex sets this field to Confirmed. If the user denies the intent, Amazon Lex sets this value to Denied. The possible values are:

Confirmed - The user has responded "Yes" to the confirmation prompt, confirming that the intent is complete and that it is ready to be fulfilled.

Denied - The user has responded "No" to the confirmation prompt.

None - The user has never been prompted for confirmation; or, the user was prompted but did not confirm or deny the prompt.

$sel:slotToElicit:IntentSummary', intentSummary_slotToElicit - The next slot to elicit from the user. If there is not slot to elicit, the field is blank.

$sel:dialogActionType:IntentSummary', intentSummary_dialogActionType - The next action that the bot should take in its interaction with the user. The possible values are:

ConfirmIntent - The next action is asking the user if the intent is complete and ready to be fulfilled. This is a yes/no question such as "Place the order?"

Close - Indicates that the there will not be a response from the user. For example, the statement "Your order has been placed" does not require a response.

ElicitIntent - The next action is to determine the intent that the user wants to fulfill.

ElicitSlot - The next action is to elicit a slot value from the user.

intentSummary_checkpointLabel :: Lens' IntentSummary (Maybe Text)

A user-defined label that identifies a particular intent. You can use this label to return to a previous intent.

Use the checkpointLabelFilter parameter of the GetSessionRequest operation to filter the intents returned by the operation to those with only the specified label.

intentSummary_slots :: Lens' IntentSummary (Maybe (HashMap Text Text))

Map of the slots that have been gathered and their values.

intentSummary_intentName :: Lens' IntentSummary (Maybe Text)

The name of the intent.

intentSummary_fulfillmentState :: Lens' IntentSummary (Maybe FulfillmentState)

The fulfillment state of the intent. The possible values are:

Failed - The Lambda function associated with the intent failed to fulfill the intent.

Fulfilled - The intent has fulfilled by the Lambda function associated with the intent.

ReadyForFulfillment - All of the information necessary for the intent is present and the intent ready to be fulfilled by the client application.

intentSummary_confirmationStatus :: Lens' IntentSummary (Maybe ConfirmationStatus)

The status of the intent after the user responds to the confirmation prompt. If the user confirms the intent, Amazon Lex sets this field to Confirmed. If the user denies the intent, Amazon Lex sets this value to Denied. The possible values are:

Confirmed - The user has responded "Yes" to the confirmation prompt, confirming that the intent is complete and that it is ready to be fulfilled.

Denied - The user has responded "No" to the confirmation prompt.

None - The user has never been prompted for confirmation; or, the user was prompted but did not confirm or deny the prompt.

intentSummary_slotToElicit :: Lens' IntentSummary (Maybe Text)

The next slot to elicit from the user. If there is not slot to elicit, the field is blank.

intentSummary_dialogActionType :: Lens' IntentSummary DialogActionType

The next action that the bot should take in its interaction with the user. The possible values are:

ConfirmIntent - The next action is asking the user if the intent is complete and ready to be fulfilled. This is a yes/no question such as "Place the order?"

Close - Indicates that the there will not be a response from the user. For example, the statement "Your order has been placed" does not require a response.

ElicitIntent - The next action is to determine the intent that the user wants to fulfill.

ElicitSlot - The next action is to elicit a slot value from the user.

PredictedIntent

data PredictedIntent

An intent that Amazon Lex suggests satisfies the user's intent. Includes the name of the intent, the confidence that Amazon Lex has that the user's intent is satisfied, and the slots defined for the intent.

See: newPredictedIntent smart constructor.

Constructors

PredictedIntent'

Fields

nluIntentConfidence :: Maybe IntentConfidence

Indicates how confident Amazon Lex is that an intent satisfies the user's intent.

slots :: Maybe (Sensitive (HashMap Text Text))

The slot and slot values associated with the predicted intent.

intentName :: Maybe Text

The name of the intent that Amazon Lex suggests satisfies the user's intent.

newPredictedIntent :: PredictedIntent

Create a value of PredictedIntent with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:nluIntentConfidence:PredictedIntent', predictedIntent_nluIntentConfidence - Indicates how confident Amazon Lex is that an intent satisfies the user's intent.

$sel:slots:PredictedIntent', predictedIntent_slots - The slot and slot values associated with the predicted intent.

$sel:intentName:PredictedIntent', predictedIntent_intentName - The name of the intent that Amazon Lex suggests satisfies the user's intent.

predictedIntent_nluIntentConfidence :: Lens' PredictedIntent (Maybe IntentConfidence)

Indicates how confident Amazon Lex is that an intent satisfies the user's intent.

predictedIntent_slots :: Lens' PredictedIntent (Maybe (HashMap Text Text))

The slot and slot values associated with the predicted intent.

predictedIntent_intentName :: Lens' PredictedIntent (Maybe Text)

The name of the intent that Amazon Lex suggests satisfies the user's intent.

ResponseCard

data ResponseCard

If you configure a response card when creating your bots, Amazon Lex substitutes the session attributes and slot values that are available, and then returns it. The response card can also come from a Lambda function ( dialogCodeHook and fulfillmentActivity on an intent).

See: newResponseCard smart constructor.

Constructors

ResponseCard'

Fields

genericAttachments :: Maybe [GenericAttachment]

An array of attachment objects representing options.

version :: Maybe Text

The version of the response card format.

contentType :: Maybe ContentType

The content type of the response.

newResponseCard :: ResponseCard

Create a value of ResponseCard with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:genericAttachments:ResponseCard', responseCard_genericAttachments - An array of attachment objects representing options.

$sel:version:ResponseCard', responseCard_version - The version of the response card format.

$sel:contentType:ResponseCard', responseCard_contentType - The content type of the response.

responseCard_genericAttachments :: Lens' ResponseCard (Maybe [GenericAttachment])

An array of attachment objects representing options.

responseCard_version :: Lens' ResponseCard (Maybe Text)

The version of the response card format.

responseCard_contentType :: Lens' ResponseCard (Maybe ContentType)

The content type of the response.

SentimentResponse

data SentimentResponse

The sentiment expressed in an utterance.

When the bot is configured to send utterances to Amazon Comprehend for sentiment analysis, this field structure contains the result of the analysis.

See: newSentimentResponse smart constructor.

Constructors

SentimentResponse'

Fields

sentimentScore :: Maybe Text

The likelihood that the sentiment was correctly inferred.

sentimentLabel :: Maybe Text

The inferred sentiment that Amazon Comprehend has the highest confidence in.

newSentimentResponse :: SentimentResponse

Create a value of SentimentResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:sentimentScore:SentimentResponse', sentimentResponse_sentimentScore - The likelihood that the sentiment was correctly inferred.

$sel:sentimentLabel:SentimentResponse', sentimentResponse_sentimentLabel - The inferred sentiment that Amazon Comprehend has the highest confidence in.

sentimentResponse_sentimentScore :: Lens' SentimentResponse (Maybe Text)

The likelihood that the sentiment was correctly inferred.

sentimentResponse_sentimentLabel :: Lens' SentimentResponse (Maybe Text)

The inferred sentiment that Amazon Comprehend has the highest confidence in.