Amazonka.LexRuntime.Types.DialogAction

Description

Documentation

data DialogAction

Describes the next action that the bot should take in its interaction with the user and provides information about the context in which the action takes place. Use the DialogAction data type to set the interaction to a specific state, or to return the interaction to a previous state.

See: newDialogAction smart constructor.

Constructors

DialogAction'

Fields

slots :: Maybe (Sensitive (HashMap Text Text))

Map of the slots that have been gathered and their values.

intentName :: Maybe Text

The name of the intent.

fulfillmentState :: Maybe FulfillmentState

The fulfillment state of the intent. The possible values are:

Failed - The Lambda function associated with the intent failed to fulfill the intent.

Fulfilled - The intent has fulfilled by the Lambda function associated with the intent.

ReadyForFulfillment - All of the information necessary for the intent is present and the intent ready to be fulfilled by the client application.

messageFormat :: Maybe MessageFormatType

PlainText - The message contains plain UTF-8 text.

CustomPayload - The message is a custom format for the client.

SSML - The message contains text formatted for voice output.

Composite - The message contains an escaped JSON object containing one or more messages. For more information, see Message Groups.

message :: Maybe (Sensitive Text)

The message that should be shown to the user. If you don't specify a message, Amazon Lex will use the message configured for the intent.

slotToElicit :: Maybe Text

The name of the slot that should be elicited from the user.

type' :: DialogActionType

The next action that the bot should take in its interaction with the user. The possible values are:

ConfirmIntent - The next action is asking the user if the intent is complete and ready to be fulfilled. This is a yes/no question such as "Place the order?"

Close - Indicates that the there will not be a response from the user. For example, the statement "Your order has been placed" does not require a response.

Delegate - The next action is determined by Amazon Lex.

ElicitIntent - The next action is to determine the intent that the user wants to fulfill.

ElicitSlot - The next action is to elicit a slot value from the user.

newDialogAction

Arguments

:: DialogActionType
$sel:type':DialogAction'

-> DialogAction

Create a value of DialogAction with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:slots:DialogAction', dialogAction_slots - Map of the slots that have been gathered and their values.

$sel:intentName:DialogAction', dialogAction_intentName - The name of the intent.

$sel:fulfillmentState:DialogAction', dialogAction_fulfillmentState - The fulfillment state of the intent. The possible values are:

Failed - The Lambda function associated with the intent failed to fulfill the intent.

Fulfilled - The intent has fulfilled by the Lambda function associated with the intent.

ReadyForFulfillment - All of the information necessary for the intent is present and the intent ready to be fulfilled by the client application.

$sel:messageFormat:DialogAction', dialogAction_messageFormat - - PlainText - The message contains plain UTF-8 text.

CustomPayload - The message is a custom format for the client.

SSML - The message contains text formatted for voice output.

Composite - The message contains an escaped JSON object containing one or more messages. For more information, see Message Groups.

$sel:message:DialogAction', dialogAction_message - The message that should be shown to the user. If you don't specify a message, Amazon Lex will use the message configured for the intent.

$sel:slotToElicit:DialogAction', dialogAction_slotToElicit - The name of the slot that should be elicited from the user.

$sel:type':DialogAction', dialogAction_type - The next action that the bot should take in its interaction with the user. The possible values are:

ConfirmIntent - The next action is asking the user if the intent is complete and ready to be fulfilled. This is a yes/no question such as "Place the order?"

Close - Indicates that the there will not be a response from the user. For example, the statement "Your order has been placed" does not require a response.

Delegate - The next action is determined by Amazon Lex.

ElicitIntent - The next action is to determine the intent that the user wants to fulfill.

ElicitSlot - The next action is to elicit a slot value from the user.

dialogAction_slots :: Lens' DialogAction (Maybe (HashMap Text Text))

Map of the slots that have been gathered and their values.

dialogAction_intentName :: Lens' DialogAction (Maybe Text)

The name of the intent.

dialogAction_fulfillmentState :: Lens' DialogAction (Maybe FulfillmentState)

The fulfillment state of the intent. The possible values are:

Failed - The Lambda function associated with the intent failed to fulfill the intent.

Fulfilled - The intent has fulfilled by the Lambda function associated with the intent.

ReadyForFulfillment - All of the information necessary for the intent is present and the intent ready to be fulfilled by the client application.

dialogAction_messageFormat :: Lens' DialogAction (Maybe MessageFormatType)

PlainText - The message contains plain UTF-8 text.

CustomPayload - The message is a custom format for the client.

SSML - The message contains text formatted for voice output.

Composite - The message contains an escaped JSON object containing one or more messages. For more information, see Message Groups.

dialogAction_message :: Lens' DialogAction (Maybe Text)

The message that should be shown to the user. If you don't specify a message, Amazon Lex will use the message configured for the intent.

dialogAction_slotToElicit :: Lens' DialogAction (Maybe Text)

The name of the slot that should be elicited from the user.

dialogAction_type :: Lens' DialogAction DialogActionType

The next action that the bot should take in its interaction with the user. The possible values are:

ConfirmIntent - The next action is asking the user if the intent is complete and ready to be fulfilled. This is a yes/no question such as "Place the order?"

Close - Indicates that the there will not be a response from the user. For example, the statement "Your order has been placed" does not require a response.

Delegate - The next action is determined by Amazon Lex.

ElicitIntent - The next action is to determine the intent that the user wants to fulfill.

ElicitSlot - The next action is to elicit a slot value from the user.