Amazonka.LicenseManager.CheckInLicense

Description

Checks in the specified license. Check in a license when it is no longer in use.

Creating a Request

data CheckInLicense

See: newCheckInLicense smart constructor.

Constructors

CheckInLicense'

Fields

beneficiary :: Maybe Text

License beneficiary.

licenseConsumptionToken :: Text

License consumption token.

newCheckInLicense

Arguments

:: Text
$sel:licenseConsumptionToken:CheckInLicense'

-> CheckInLicense

Create a value of CheckInLicense with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:beneficiary:CheckInLicense', checkInLicense_beneficiary - License beneficiary.

$sel:licenseConsumptionToken:CheckInLicense', checkInLicense_licenseConsumptionToken - License consumption token.

Request Lenses

checkInLicense_beneficiary :: Lens' CheckInLicense (Maybe Text)

License beneficiary.

checkInLicense_licenseConsumptionToken :: Lens' CheckInLicense Text

License consumption token.

Destructuring the Response

data CheckInLicenseResponse

See: newCheckInLicenseResponse smart constructor.

Constructors

CheckInLicenseResponse'

Fields

httpStatus :: Int

The response's http status code.

newCheckInLicenseResponse

Arguments

:: Int
$sel:httpStatus:CheckInLicenseResponse'

-> CheckInLicenseResponse

Create a value of CheckInLicenseResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:httpStatus:CheckInLicenseResponse', checkInLicenseResponse_httpStatus - The response's http status code.

Response Lenses

checkInLicenseResponse_httpStatus :: Lens' CheckInLicenseResponse Int

The response's http status code.