Amazonka.LicenseManager.CheckoutBorrowLicense

Description

Checks out the specified license for offline use.

Creating a Request

data CheckoutBorrowLicense

See: newCheckoutBorrowLicense smart constructor.

Constructors

CheckoutBorrowLicense'

Fields

checkoutMetadata :: Maybe [Metadata]

Information about constraints.

nodeId :: Maybe Text

Node ID.

licenseArn :: Text

Amazon Resource Name (ARN) of the license. The license must use the borrow consumption configuration.

entitlements :: [EntitlementData]

License entitlements. Partial checkouts are not supported.

digitalSignatureMethod :: DigitalSignatureMethod

Digital signature method. The possible value is JSON Web Signature (JWS) algorithm PS384. For more information, see RFC 7518 Digital Signature with RSASSA-PSS.

clientToken :: Text

Unique, case-sensitive identifier that you provide to ensure the idempotency of the request.

newCheckoutBorrowLicense

Arguments

:: Text
$sel:licenseArn:CheckoutBorrowLicense'

-> DigitalSignatureMethod
$sel:digitalSignatureMethod:CheckoutBorrowLicense'

-> Text
$sel:clientToken:CheckoutBorrowLicense'

-> CheckoutBorrowLicense

Create a value of CheckoutBorrowLicense with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:checkoutMetadata:CheckoutBorrowLicense', checkoutBorrowLicense_checkoutMetadata - Information about constraints.

$sel:nodeId:CheckoutBorrowLicense', checkoutBorrowLicense_nodeId - Node ID.

$sel:licenseArn:CheckoutBorrowLicense', checkoutBorrowLicense_licenseArn - Amazon Resource Name (ARN) of the license. The license must use the borrow consumption configuration.

$sel:entitlements:CheckoutBorrowLicense', checkoutBorrowLicense_entitlements - License entitlements. Partial checkouts are not supported.

$sel:digitalSignatureMethod:CheckoutBorrowLicense', checkoutBorrowLicense_digitalSignatureMethod - Digital signature method. The possible value is JSON Web Signature (JWS) algorithm PS384. For more information, see RFC 7518 Digital Signature with RSASSA-PSS.

$sel:clientToken:CheckoutBorrowLicense', checkoutBorrowLicense_clientToken - Unique, case-sensitive identifier that you provide to ensure the idempotency of the request.

Request Lenses

checkoutBorrowLicense_checkoutMetadata :: Lens' CheckoutBorrowLicense (Maybe [Metadata])

Information about constraints.

checkoutBorrowLicense_nodeId :: Lens' CheckoutBorrowLicense (Maybe Text)

Node ID.

checkoutBorrowLicense_licenseArn :: Lens' CheckoutBorrowLicense Text

Amazon Resource Name (ARN) of the license. The license must use the borrow consumption configuration.

checkoutBorrowLicense_entitlements :: Lens' CheckoutBorrowLicense [EntitlementData]

License entitlements. Partial checkouts are not supported.

checkoutBorrowLicense_digitalSignatureMethod :: Lens' CheckoutBorrowLicense DigitalSignatureMethod

Digital signature method. The possible value is JSON Web Signature (JWS) algorithm PS384. For more information, see RFC 7518 Digital Signature with RSASSA-PSS.

checkoutBorrowLicense_clientToken :: Lens' CheckoutBorrowLicense Text

Unique, case-sensitive identifier that you provide to ensure the idempotency of the request.

Destructuring the Response

data CheckoutBorrowLicenseResponse

See: newCheckoutBorrowLicenseResponse smart constructor.

Constructors

CheckoutBorrowLicenseResponse'

Fields

entitlementsAllowed :: Maybe [EntitlementData]

Allowed license entitlements.

checkoutMetadata :: Maybe [Metadata]

Information about constraints.

expiration :: Maybe Text

Date and time at which the license checkout expires.

signedToken :: Maybe Text

Signed token.

licenseConsumptionToken :: Maybe Text

License consumption token.

nodeId :: Maybe Text

Node ID.

issuedAt :: Maybe Text

Date and time at which the license checkout is issued.

licenseArn :: Maybe Text

Amazon Resource Name (ARN) of the license.

httpStatus :: Int

The response's http status code.

newCheckoutBorrowLicenseResponse

Arguments

:: Int
$sel:httpStatus:CheckoutBorrowLicenseResponse'

-> CheckoutBorrowLicenseResponse

Create a value of CheckoutBorrowLicenseResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:entitlementsAllowed:CheckoutBorrowLicenseResponse', checkoutBorrowLicenseResponse_entitlementsAllowed - Allowed license entitlements.

$sel:checkoutMetadata:CheckoutBorrowLicense', checkoutBorrowLicenseResponse_checkoutMetadata - Information about constraints.

$sel:expiration:CheckoutBorrowLicenseResponse', checkoutBorrowLicenseResponse_expiration - Date and time at which the license checkout expires.

$sel:signedToken:CheckoutBorrowLicenseResponse', checkoutBorrowLicenseResponse_signedToken - Signed token.

$sel:licenseConsumptionToken:CheckoutBorrowLicenseResponse', checkoutBorrowLicenseResponse_licenseConsumptionToken - License consumption token.

$sel:nodeId:CheckoutBorrowLicense', checkoutBorrowLicenseResponse_nodeId - Node ID.

$sel:issuedAt:CheckoutBorrowLicenseResponse', checkoutBorrowLicenseResponse_issuedAt - Date and time at which the license checkout is issued.

$sel:licenseArn:CheckoutBorrowLicense', checkoutBorrowLicenseResponse_licenseArn - Amazon Resource Name (ARN) of the license.

$sel:httpStatus:CheckoutBorrowLicenseResponse', checkoutBorrowLicenseResponse_httpStatus - The response's http status code.

Response Lenses

checkoutBorrowLicenseResponse_entitlementsAllowed :: Lens' CheckoutBorrowLicenseResponse (Maybe [EntitlementData])

Allowed license entitlements.

checkoutBorrowLicenseResponse_checkoutMetadata :: Lens' CheckoutBorrowLicenseResponse (Maybe [Metadata])

Information about constraints.

checkoutBorrowLicenseResponse_expiration :: Lens' CheckoutBorrowLicenseResponse (Maybe Text)

Date and time at which the license checkout expires.

checkoutBorrowLicenseResponse_signedToken :: Lens' CheckoutBorrowLicenseResponse (Maybe Text)

Signed token.

checkoutBorrowLicenseResponse_licenseConsumptionToken :: Lens' CheckoutBorrowLicenseResponse (Maybe Text)

License consumption token.

checkoutBorrowLicenseResponse_nodeId :: Lens' CheckoutBorrowLicenseResponse (Maybe Text)

Node ID.

checkoutBorrowLicenseResponse_issuedAt :: Lens' CheckoutBorrowLicenseResponse (Maybe Text)

Date and time at which the license checkout is issued.

checkoutBorrowLicenseResponse_licenseArn :: Lens' CheckoutBorrowLicenseResponse (Maybe Text)

Amazon Resource Name (ARN) of the license.

checkoutBorrowLicenseResponse_httpStatus :: Lens' CheckoutBorrowLicenseResponse Int

The response's http status code.