Amazonka.LicenseManager.CreateGrant

Description

Creates a grant for the specified license. A grant shares the use of license entitlements with specific Amazon Web Services accounts.

Creating a Request

data CreateGrant

See: newCreateGrant smart constructor.

Constructors

CreateGrant'

Fields

clientToken :: Text

Unique, case-sensitive identifier that you provide to ensure the idempotency of the request.

grantName :: Text

Grant name.

licenseArn :: Text

Amazon Resource Name (ARN) of the license.

principals :: NonEmpty Text

The grant principals.

homeRegion :: Text

Home Region of the grant.

allowedOperations :: NonEmpty AllowedOperation

Allowed operations for the grant.

newCreateGrant

Arguments

:: Text
$sel:clientToken:CreateGrant'

-> Text
$sel:grantName:CreateGrant'

-> Text
$sel:licenseArn:CreateGrant'

-> NonEmpty Text
$sel:principals:CreateGrant'

-> Text
$sel:homeRegion:CreateGrant'

-> NonEmpty AllowedOperation
$sel:allowedOperations:CreateGrant'

-> CreateGrant

Create a value of CreateGrant with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:clientToken:CreateGrant', createGrant_clientToken - Unique, case-sensitive identifier that you provide to ensure the idempotency of the request.

$sel:grantName:CreateGrant', createGrant_grantName - Grant name.

$sel:licenseArn:CreateGrant', createGrant_licenseArn - Amazon Resource Name (ARN) of the license.

$sel:principals:CreateGrant', createGrant_principals - The grant principals.

$sel:homeRegion:CreateGrant', createGrant_homeRegion - Home Region of the grant.

$sel:allowedOperations:CreateGrant', createGrant_allowedOperations - Allowed operations for the grant.

Request Lenses

createGrant_clientToken :: Lens' CreateGrant Text

Unique, case-sensitive identifier that you provide to ensure the idempotency of the request.

createGrant_grantName :: Lens' CreateGrant Text

Grant name.

createGrant_licenseArn :: Lens' CreateGrant Text

Amazon Resource Name (ARN) of the license.

createGrant_principals :: Lens' CreateGrant (NonEmpty Text)

The grant principals.

createGrant_homeRegion :: Lens' CreateGrant Text

Home Region of the grant.

createGrant_allowedOperations :: Lens' CreateGrant (NonEmpty AllowedOperation)

Allowed operations for the grant.

Destructuring the Response

data CreateGrantResponse

See: newCreateGrantResponse smart constructor.

Constructors

CreateGrantResponse'

Fields

status :: Maybe GrantStatus

Grant status.

version :: Maybe Text

Grant version.

grantArn :: Maybe Text

Grant ARN.

httpStatus :: Int

The response's http status code.

newCreateGrantResponse

Arguments

:: Int
$sel:httpStatus:CreateGrantResponse'

-> CreateGrantResponse

Create a value of CreateGrantResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:status:CreateGrantResponse', createGrantResponse_status - Grant status.

$sel:version:CreateGrantResponse', createGrantResponse_version - Grant version.

$sel:grantArn:CreateGrantResponse', createGrantResponse_grantArn - Grant ARN.

$sel:httpStatus:CreateGrantResponse', createGrantResponse_httpStatus - The response's http status code.

Response Lenses

createGrantResponse_status :: Lens' CreateGrantResponse (Maybe GrantStatus)

Grant status.

createGrantResponse_version :: Lens' CreateGrantResponse (Maybe Text)

Grant version.

createGrantResponse_grantArn :: Lens' CreateGrantResponse (Maybe Text)

Grant ARN.

createGrantResponse_httpStatus :: Lens' CreateGrantResponse Int

The response's http status code.