Amazonka.LicenseManager.CreateLicense

Description

Creates a license.

Creating a Request

data CreateLicense

See: newCreateLicense smart constructor.

Constructors

CreateLicense'

Fields

licenseMetadata :: Maybe [Metadata]

Information about the license.

licenseName :: Text

License name.

productName :: Text

Product name.

productSKU :: Text

Product SKU.

issuer :: Issuer

License issuer.

homeRegion :: Text

Home Region for the license.

validity :: DatetimeRange

Date and time range during which the license is valid, in ISO8601-UTC format.

entitlements :: [Entitlement]

License entitlements.

beneficiary :: Text

License beneficiary.

consumptionConfiguration :: ConsumptionConfiguration

Configuration for consumption of the license. Choose a provisional configuration for workloads running with continuous connectivity. Choose a borrow configuration for workloads with offline usage.

clientToken :: Text

Unique, case-sensitive identifier that you provide to ensure the idempotency of the request.

newCreateLicense

Arguments

:: Text
$sel:licenseName:CreateLicense'

-> Text
$sel:productName:CreateLicense'

-> Text
$sel:productSKU:CreateLicense'

-> Issuer
$sel:issuer:CreateLicense'

-> Text
$sel:homeRegion:CreateLicense'

-> DatetimeRange
$sel:validity:CreateLicense'

-> Text
$sel:beneficiary:CreateLicense'

-> ConsumptionConfiguration
$sel:consumptionConfiguration:CreateLicense'

-> Text
$sel:clientToken:CreateLicense'

-> CreateLicense

Create a value of CreateLicense with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:licenseMetadata:CreateLicense', createLicense_licenseMetadata - Information about the license.

$sel:licenseName:CreateLicense', createLicense_licenseName - License name.

$sel:productName:CreateLicense', createLicense_productName - Product name.

$sel:productSKU:CreateLicense', createLicense_productSKU - Product SKU.

$sel:issuer:CreateLicense', createLicense_issuer - License issuer.

$sel:homeRegion:CreateLicense', createLicense_homeRegion - Home Region for the license.

$sel:validity:CreateLicense', createLicense_validity - Date and time range during which the license is valid, in ISO8601-UTC format.

$sel:entitlements:CreateLicense', createLicense_entitlements - License entitlements.

$sel:beneficiary:CreateLicense', createLicense_beneficiary - License beneficiary.

$sel:consumptionConfiguration:CreateLicense', createLicense_consumptionConfiguration - Configuration for consumption of the license. Choose a provisional configuration for workloads running with continuous connectivity. Choose a borrow configuration for workloads with offline usage.

$sel:clientToken:CreateLicense', createLicense_clientToken - Unique, case-sensitive identifier that you provide to ensure the idempotency of the request.

Request Lenses

createLicense_licenseMetadata :: Lens' CreateLicense (Maybe [Metadata])

Information about the license.

createLicense_licenseName :: Lens' CreateLicense Text

License name.

createLicense_productName :: Lens' CreateLicense Text

Product name.

createLicense_productSKU :: Lens' CreateLicense Text

Product SKU.

createLicense_issuer :: Lens' CreateLicense Issuer

License issuer.

createLicense_homeRegion :: Lens' CreateLicense Text

Home Region for the license.

createLicense_validity :: Lens' CreateLicense DatetimeRange

Date and time range during which the license is valid, in ISO8601-UTC format.

createLicense_entitlements :: Lens' CreateLicense [Entitlement]

License entitlements.

createLicense_beneficiary :: Lens' CreateLicense Text

License beneficiary.

createLicense_consumptionConfiguration :: Lens' CreateLicense ConsumptionConfiguration

Configuration for consumption of the license. Choose a provisional configuration for workloads running with continuous connectivity. Choose a borrow configuration for workloads with offline usage.

createLicense_clientToken :: Lens' CreateLicense Text

Unique, case-sensitive identifier that you provide to ensure the idempotency of the request.

Destructuring the Response

data CreateLicenseResponse

See: newCreateLicenseResponse smart constructor.

Constructors

CreateLicenseResponse'

Fields

status :: Maybe LicenseStatus

License status.

version :: Maybe Text

License version.

licenseArn :: Maybe Text

Amazon Resource Name (ARN) of the license.

httpStatus :: Int

The response's http status code.

newCreateLicenseResponse

Arguments

:: Int
$sel:httpStatus:CreateLicenseResponse'

-> CreateLicenseResponse

Create a value of CreateLicenseResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:status:CreateLicenseResponse', createLicenseResponse_status - License status.

$sel:version:CreateLicenseResponse', createLicenseResponse_version - License version.

$sel:licenseArn:CreateLicenseResponse', createLicenseResponse_licenseArn - Amazon Resource Name (ARN) of the license.

$sel:httpStatus:CreateLicenseResponse', createLicenseResponse_httpStatus - The response's http status code.

Response Lenses

createLicenseResponse_status :: Lens' CreateLicenseResponse (Maybe LicenseStatus)

License status.

createLicenseResponse_version :: Lens' CreateLicenseResponse (Maybe Text)

License version.

createLicenseResponse_licenseArn :: Lens' CreateLicenseResponse (Maybe Text)

Amazon Resource Name (ARN) of the license.

createLicenseResponse_httpStatus :: Lens' CreateLicenseResponse Int

The response's http status code.