Amazonka.LicenseManager.CreateLicenseConfiguration

Description

Creates a license configuration.

A license configuration is an abstraction of a customer license agreement that can be consumed and enforced by License Manager. Components include specifications for the license type (licensing by instance, socket, CPU, or vCPU), allowed tenancy (shared tenancy, Dedicated Instance, Dedicated Host, or all of these), license affinity to host (how long a license must be associated with a host), and the number of licenses purchased and used.

Creating a Request

data CreateLicenseConfiguration

See: newCreateLicenseConfiguration smart constructor.

Constructors

CreateLicenseConfiguration'

Fields

licenseCount :: Maybe Integer

Number of licenses managed by the license configuration.

licenseCountHardLimit :: Maybe Bool

Indicates whether hard or soft license enforcement is used. Exceeding a hard limit blocks the launch of new instances.

disassociateWhenNotFound :: Maybe Bool

When true, disassociates a resource when software is uninstalled.

productInformationList :: Maybe [ProductInformation]

Product information.

licenseRules :: Maybe [Text]

License rules. The syntax is #name=value (for example, #allowedTenancy=EC2-DedicatedHost). The available rules vary by dimension, as follows.

Cores dimension: allowedTenancy | licenseAffinityToHost | maximumCores | minimumCores

Instances dimension: allowedTenancy | maximumCores | minimumCores | maximumSockets | minimumSockets | maximumVcpus | minimumVcpus

Sockets dimension: allowedTenancy | licenseAffinityToHost | maximumSockets | minimumSockets

vCPUs dimension: allowedTenancy | honorVcpuOptimization | maximumVcpus | minimumVcpus

The unit for licenseAffinityToHost is days and the range is 1 to 180. The possible values for allowedTenancy are EC2-Default, EC2-DedicatedHost, and EC2-DedicatedInstance. The possible values for honorVcpuOptimization are True and False.

description :: Maybe Text

Description of the license configuration.

tags :: Maybe [Tag]

Tags to add to the license configuration.

name :: Text

Name of the license configuration.

licenseCountingType :: LicenseCountingType

Dimension used to track the license inventory.

newCreateLicenseConfiguration

Arguments

:: Text
$sel:name:CreateLicenseConfiguration'

-> LicenseCountingType
$sel:licenseCountingType:CreateLicenseConfiguration'

-> CreateLicenseConfiguration

Create a value of CreateLicenseConfiguration with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:licenseCount:CreateLicenseConfiguration', createLicenseConfiguration_licenseCount - Number of licenses managed by the license configuration.

$sel:licenseCountHardLimit:CreateLicenseConfiguration', createLicenseConfiguration_licenseCountHardLimit - Indicates whether hard or soft license enforcement is used. Exceeding a hard limit blocks the launch of new instances.

$sel:disassociateWhenNotFound:CreateLicenseConfiguration', createLicenseConfiguration_disassociateWhenNotFound - When true, disassociates a resource when software is uninstalled.

$sel:productInformationList:CreateLicenseConfiguration', createLicenseConfiguration_productInformationList - Product information.

$sel:licenseRules:CreateLicenseConfiguration', createLicenseConfiguration_licenseRules - License rules. The syntax is #name=value (for example, #allowedTenancy=EC2-DedicatedHost). The available rules vary by dimension, as follows.

Cores dimension: allowedTenancy | licenseAffinityToHost | maximumCores | minimumCores

Instances dimension: allowedTenancy | maximumCores | minimumCores | maximumSockets | minimumSockets | maximumVcpus | minimumVcpus

Sockets dimension: allowedTenancy | licenseAffinityToHost | maximumSockets | minimumSockets

vCPUs dimension: allowedTenancy | honorVcpuOptimization | maximumVcpus | minimumVcpus

The unit for licenseAffinityToHost is days and the range is 1 to 180. The possible values for allowedTenancy are EC2-Default, EC2-DedicatedHost, and EC2-DedicatedInstance. The possible values for honorVcpuOptimization are True and False.

$sel:description:CreateLicenseConfiguration', createLicenseConfiguration_description - Description of the license configuration.

$sel:tags:CreateLicenseConfiguration', createLicenseConfiguration_tags - Tags to add to the license configuration.

$sel:name:CreateLicenseConfiguration', createLicenseConfiguration_name - Name of the license configuration.

$sel:licenseCountingType:CreateLicenseConfiguration', createLicenseConfiguration_licenseCountingType - Dimension used to track the license inventory.

Request Lenses

createLicenseConfiguration_licenseCount :: Lens' CreateLicenseConfiguration (Maybe Integer)

Number of licenses managed by the license configuration.

createLicenseConfiguration_licenseCountHardLimit :: Lens' CreateLicenseConfiguration (Maybe Bool)

Indicates whether hard or soft license enforcement is used. Exceeding a hard limit blocks the launch of new instances.

createLicenseConfiguration_disassociateWhenNotFound :: Lens' CreateLicenseConfiguration (Maybe Bool)

When true, disassociates a resource when software is uninstalled.

createLicenseConfiguration_productInformationList :: Lens' CreateLicenseConfiguration (Maybe [ProductInformation])

Product information.

createLicenseConfiguration_licenseRules :: Lens' CreateLicenseConfiguration (Maybe [Text])

License rules. The syntax is #name=value (for example, #allowedTenancy=EC2-DedicatedHost). The available rules vary by dimension, as follows.

Cores dimension: allowedTenancy | licenseAffinityToHost | maximumCores | minimumCores

Instances dimension: allowedTenancy | maximumCores | minimumCores | maximumSockets | minimumSockets | maximumVcpus | minimumVcpus

Sockets dimension: allowedTenancy | licenseAffinityToHost | maximumSockets | minimumSockets

vCPUs dimension: allowedTenancy | honorVcpuOptimization | maximumVcpus | minimumVcpus

The unit for licenseAffinityToHost is days and the range is 1 to 180. The possible values for allowedTenancy are EC2-Default, EC2-DedicatedHost, and EC2-DedicatedInstance. The possible values for honorVcpuOptimization are True and False.

createLicenseConfiguration_description :: Lens' CreateLicenseConfiguration (Maybe Text)

Description of the license configuration.

createLicenseConfiguration_tags :: Lens' CreateLicenseConfiguration (Maybe [Tag])

Tags to add to the license configuration.

createLicenseConfiguration_name :: Lens' CreateLicenseConfiguration Text

Name of the license configuration.

createLicenseConfiguration_licenseCountingType :: Lens' CreateLicenseConfiguration LicenseCountingType

Dimension used to track the license inventory.

Destructuring the Response

data CreateLicenseConfigurationResponse

See: newCreateLicenseConfigurationResponse smart constructor.

Constructors

CreateLicenseConfigurationResponse'

Fields

licenseConfigurationArn :: Maybe Text

Amazon Resource Name (ARN) of the license configuration.

httpStatus :: Int

The response's http status code.

newCreateLicenseConfigurationResponse

Arguments

:: Int
$sel:httpStatus:CreateLicenseConfigurationResponse'

-> CreateLicenseConfigurationResponse

Create a value of CreateLicenseConfigurationResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:licenseConfigurationArn:CreateLicenseConfigurationResponse', createLicenseConfigurationResponse_licenseConfigurationArn - Amazon Resource Name (ARN) of the license configuration.

$sel:httpStatus:CreateLicenseConfigurationResponse', createLicenseConfigurationResponse_httpStatus - The response's http status code.

Response Lenses

createLicenseConfigurationResponse_licenseConfigurationArn :: Lens' CreateLicenseConfigurationResponse (Maybe Text)

Amazon Resource Name (ARN) of the license configuration.

createLicenseConfigurationResponse_httpStatus :: Lens' CreateLicenseConfigurationResponse Int

The response's http status code.