Amazonka.LicenseManager.CreateLicenseManagerReportGenerator

Description

Creates a report generator.

Creating a Request

data CreateLicenseManagerReportGenerator

See: newCreateLicenseManagerReportGenerator smart constructor.

Constructors

CreateLicenseManagerReportGenerator'

Fields

description :: Maybe Text

Description of the report generator.

tags :: Maybe [Tag]

Tags to add to the report generator.

reportGeneratorName :: Text

Name of the report generator.

type' :: [ReportType]

Type of reports to generate. The following report types an be generated:

License configuration report - Reports the number and details of consumed licenses for a license configuration.

Resource report - Reports the tracked licenses and resource consumption for a license configuration.

reportContext :: ReportContext

Defines the type of license configuration the report generator tracks.

reportFrequency :: ReportFrequency

Frequency by which reports are generated. Reports can be generated daily, monthly, or weekly.

clientToken :: Text

Unique, case-sensitive identifier that you provide to ensure the idempotency of the request.

newCreateLicenseManagerReportGenerator

Arguments

:: Text
$sel:reportGeneratorName:CreateLicenseManagerReportGenerator'

-> ReportContext
$sel:reportContext:CreateLicenseManagerReportGenerator'

-> ReportFrequency
$sel:reportFrequency:CreateLicenseManagerReportGenerator'

-> Text
$sel:clientToken:CreateLicenseManagerReportGenerator'

-> CreateLicenseManagerReportGenerator

Create a value of CreateLicenseManagerReportGenerator with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:description:CreateLicenseManagerReportGenerator', createLicenseManagerReportGenerator_description - Description of the report generator.

$sel:tags:CreateLicenseManagerReportGenerator', createLicenseManagerReportGenerator_tags - Tags to add to the report generator.

$sel:reportGeneratorName:CreateLicenseManagerReportGenerator', createLicenseManagerReportGenerator_reportGeneratorName - Name of the report generator.

$sel:type':CreateLicenseManagerReportGenerator', createLicenseManagerReportGenerator_type - Type of reports to generate. The following report types an be generated:

License configuration report - Reports the number and details of consumed licenses for a license configuration.

Resource report - Reports the tracked licenses and resource consumption for a license configuration.

$sel:reportContext:CreateLicenseManagerReportGenerator', createLicenseManagerReportGenerator_reportContext - Defines the type of license configuration the report generator tracks.

$sel:reportFrequency:CreateLicenseManagerReportGenerator', createLicenseManagerReportGenerator_reportFrequency - Frequency by which reports are generated. Reports can be generated daily, monthly, or weekly.

$sel:clientToken:CreateLicenseManagerReportGenerator', createLicenseManagerReportGenerator_clientToken - Unique, case-sensitive identifier that you provide to ensure the idempotency of the request.

Request Lenses

createLicenseManagerReportGenerator_description :: Lens' CreateLicenseManagerReportGenerator (Maybe Text)

Description of the report generator.

createLicenseManagerReportGenerator_tags :: Lens' CreateLicenseManagerReportGenerator (Maybe [Tag])

Tags to add to the report generator.

createLicenseManagerReportGenerator_reportGeneratorName :: Lens' CreateLicenseManagerReportGenerator Text

Name of the report generator.

createLicenseManagerReportGenerator_type :: Lens' CreateLicenseManagerReportGenerator [ReportType]

Type of reports to generate. The following report types an be generated:

License configuration report - Reports the number and details of consumed licenses for a license configuration.

Resource report - Reports the tracked licenses and resource consumption for a license configuration.

createLicenseManagerReportGenerator_reportContext :: Lens' CreateLicenseManagerReportGenerator ReportContext

Defines the type of license configuration the report generator tracks.

createLicenseManagerReportGenerator_reportFrequency :: Lens' CreateLicenseManagerReportGenerator ReportFrequency

Frequency by which reports are generated. Reports can be generated daily, monthly, or weekly.

createLicenseManagerReportGenerator_clientToken :: Lens' CreateLicenseManagerReportGenerator Text

Unique, case-sensitive identifier that you provide to ensure the idempotency of the request.

Destructuring the Response

data CreateLicenseManagerReportGeneratorResponse

See: newCreateLicenseManagerReportGeneratorResponse smart constructor.

Constructors

CreateLicenseManagerReportGeneratorResponse'

Fields

licenseManagerReportGeneratorArn :: Maybe Text

The Amazon Resource Name (ARN) of the new report generator.

httpStatus :: Int

The response's http status code.

newCreateLicenseManagerReportGeneratorResponse

Arguments

:: Int
$sel:httpStatus:CreateLicenseManagerReportGeneratorResponse'

-> CreateLicenseManagerReportGeneratorResponse

Create a value of CreateLicenseManagerReportGeneratorResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:licenseManagerReportGeneratorArn:CreateLicenseManagerReportGeneratorResponse', createLicenseManagerReportGeneratorResponse_licenseManagerReportGeneratorArn - The Amazon Resource Name (ARN) of the new report generator.

$sel:httpStatus:CreateLicenseManagerReportGeneratorResponse', createLicenseManagerReportGeneratorResponse_httpStatus - The response's http status code.

Response Lenses

createLicenseManagerReportGeneratorResponse_licenseManagerReportGeneratorArn :: Lens' CreateLicenseManagerReportGeneratorResponse (Maybe Text)

The Amazon Resource Name (ARN) of the new report generator.

createLicenseManagerReportGeneratorResponse_httpStatus :: Lens' CreateLicenseManagerReportGeneratorResponse Int

The response's http status code.