Amazonka.LicenseManager.CreateLicenseVersion

Description

Creates a new version of the specified license.

Creating a Request

data CreateLicenseVersion

See: newCreateLicenseVersion smart constructor.

Constructors

CreateLicenseVersion'

Fields

sourceVersion :: Maybe Text

Current version of the license.

licenseMetadata :: Maybe [Metadata]

Information about the license.

licenseArn :: Text

Amazon Resource Name (ARN) of the license.

licenseName :: Text

License name.

productName :: Text

Product name.

issuer :: Issuer

License issuer.

homeRegion :: Text

Home Region of the license.

validity :: DatetimeRange

Date and time range during which the license is valid, in ISO8601-UTC format.

entitlements :: [Entitlement]

License entitlements.

consumptionConfiguration :: ConsumptionConfiguration

Configuration for consumption of the license. Choose a provisional configuration for workloads running with continuous connectivity. Choose a borrow configuration for workloads with offline usage.

status :: LicenseStatus

License status.

clientToken :: Text

Unique, case-sensitive identifier that you provide to ensure the idempotency of the request.

newCreateLicenseVersion

Arguments

:: Text
$sel:licenseArn:CreateLicenseVersion'

-> Text
$sel:licenseName:CreateLicenseVersion'

-> Text
$sel:productName:CreateLicenseVersion'

-> Issuer
$sel:issuer:CreateLicenseVersion'

-> Text
$sel:homeRegion:CreateLicenseVersion'

-> DatetimeRange
$sel:validity:CreateLicenseVersion'

-> ConsumptionConfiguration
$sel:consumptionConfiguration:CreateLicenseVersion'

-> LicenseStatus
$sel:status:CreateLicenseVersion'

-> Text
$sel:clientToken:CreateLicenseVersion'

-> CreateLicenseVersion

Create a value of CreateLicenseVersion with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:sourceVersion:CreateLicenseVersion', createLicenseVersion_sourceVersion - Current version of the license.

$sel:licenseMetadata:CreateLicenseVersion', createLicenseVersion_licenseMetadata - Information about the license.

$sel:licenseArn:CreateLicenseVersion', createLicenseVersion_licenseArn - Amazon Resource Name (ARN) of the license.

$sel:licenseName:CreateLicenseVersion', createLicenseVersion_licenseName - License name.

$sel:productName:CreateLicenseVersion', createLicenseVersion_productName - Product name.

$sel:issuer:CreateLicenseVersion', createLicenseVersion_issuer - License issuer.

$sel:homeRegion:CreateLicenseVersion', createLicenseVersion_homeRegion - Home Region of the license.

$sel:validity:CreateLicenseVersion', createLicenseVersion_validity - Date and time range during which the license is valid, in ISO8601-UTC format.

$sel:entitlements:CreateLicenseVersion', createLicenseVersion_entitlements - License entitlements.

$sel:consumptionConfiguration:CreateLicenseVersion', createLicenseVersion_consumptionConfiguration - Configuration for consumption of the license. Choose a provisional configuration for workloads running with continuous connectivity. Choose a borrow configuration for workloads with offline usage.

$sel:status:CreateLicenseVersion', createLicenseVersion_status - License status.

$sel:clientToken:CreateLicenseVersion', createLicenseVersion_clientToken - Unique, case-sensitive identifier that you provide to ensure the idempotency of the request.

Request Lenses

createLicenseVersion_sourceVersion :: Lens' CreateLicenseVersion (Maybe Text)

Current version of the license.

createLicenseVersion_licenseMetadata :: Lens' CreateLicenseVersion (Maybe [Metadata])

Information about the license.

createLicenseVersion_licenseArn :: Lens' CreateLicenseVersion Text

Amazon Resource Name (ARN) of the license.

createLicenseVersion_licenseName :: Lens' CreateLicenseVersion Text

License name.

createLicenseVersion_productName :: Lens' CreateLicenseVersion Text

Product name.

createLicenseVersion_issuer :: Lens' CreateLicenseVersion Issuer

License issuer.

createLicenseVersion_homeRegion :: Lens' CreateLicenseVersion Text

Home Region of the license.

createLicenseVersion_validity :: Lens' CreateLicenseVersion DatetimeRange

Date and time range during which the license is valid, in ISO8601-UTC format.

createLicenseVersion_entitlements :: Lens' CreateLicenseVersion [Entitlement]

License entitlements.

createLicenseVersion_consumptionConfiguration :: Lens' CreateLicenseVersion ConsumptionConfiguration

Configuration for consumption of the license. Choose a provisional configuration for workloads running with continuous connectivity. Choose a borrow configuration for workloads with offline usage.

createLicenseVersion_status :: Lens' CreateLicenseVersion LicenseStatus

License status.

createLicenseVersion_clientToken :: Lens' CreateLicenseVersion Text

Unique, case-sensitive identifier that you provide to ensure the idempotency of the request.

Destructuring the Response

data CreateLicenseVersionResponse

See: newCreateLicenseVersionResponse smart constructor.

Constructors

CreateLicenseVersionResponse'

Fields

status :: Maybe LicenseStatus

License status.

version :: Maybe Text

New version of the license.

licenseArn :: Maybe Text

License ARN.

httpStatus :: Int

The response's http status code.

newCreateLicenseVersionResponse

Arguments

:: Int
$sel:httpStatus:CreateLicenseVersionResponse'

-> CreateLicenseVersionResponse

Create a value of CreateLicenseVersionResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:status:CreateLicenseVersion', createLicenseVersionResponse_status - License status.

$sel:version:CreateLicenseVersionResponse', createLicenseVersionResponse_version - New version of the license.

$sel:licenseArn:CreateLicenseVersion', createLicenseVersionResponse_licenseArn - License ARN.

$sel:httpStatus:CreateLicenseVersionResponse', createLicenseVersionResponse_httpStatus - The response's http status code.

Response Lenses

createLicenseVersionResponse_status :: Lens' CreateLicenseVersionResponse (Maybe LicenseStatus)

License status.

createLicenseVersionResponse_version :: Lens' CreateLicenseVersionResponse (Maybe Text)

New version of the license.

createLicenseVersionResponse_licenseArn :: Lens' CreateLicenseVersionResponse (Maybe Text)

License ARN.

createLicenseVersionResponse_httpStatus :: Lens' CreateLicenseVersionResponse Int

The response's http status code.