Amazonka.LicenseManager.DeleteLicense

Description

Deletes the specified license.

Creating a Request

data DeleteLicense

See: newDeleteLicense smart constructor.

Constructors

DeleteLicense'

Fields

licenseArn :: Text

Amazon Resource Name (ARN) of the license.

sourceVersion :: Text

Current version of the license.

newDeleteLicense

Arguments

:: Text
$sel:licenseArn:DeleteLicense'

-> Text
$sel:sourceVersion:DeleteLicense'

-> DeleteLicense

Create a value of DeleteLicense with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:licenseArn:DeleteLicense', deleteLicense_licenseArn - Amazon Resource Name (ARN) of the license.

$sel:sourceVersion:DeleteLicense', deleteLicense_sourceVersion - Current version of the license.

Request Lenses

deleteLicense_licenseArn :: Lens' DeleteLicense Text

Amazon Resource Name (ARN) of the license.

deleteLicense_sourceVersion :: Lens' DeleteLicense Text

Current version of the license.

Destructuring the Response

data DeleteLicenseResponse

See: newDeleteLicenseResponse smart constructor.

Constructors

DeleteLicenseResponse'

Fields

status :: Maybe LicenseDeletionStatus

License status.

deletionDate :: Maybe Text

Date when the license is deleted.

httpStatus :: Int

The response's http status code.

newDeleteLicenseResponse

Arguments

:: Int
$sel:httpStatus:DeleteLicenseResponse'

-> DeleteLicenseResponse

Create a value of DeleteLicenseResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:status:DeleteLicenseResponse', deleteLicenseResponse_status - License status.

$sel:deletionDate:DeleteLicenseResponse', deleteLicenseResponse_deletionDate - Date when the license is deleted.

$sel:httpStatus:DeleteLicenseResponse', deleteLicenseResponse_httpStatus - The response's http status code.

Response Lenses

deleteLicenseResponse_status :: Lens' DeleteLicenseResponse (Maybe LicenseDeletionStatus)

License status.

deleteLicenseResponse_deletionDate :: Lens' DeleteLicenseResponse (Maybe Text)

Date when the license is deleted.

deleteLicenseResponse_httpStatus :: Lens' DeleteLicenseResponse Int

The response's http status code.