Amazonka.LicenseManager.GetLicenseUsage

Description

Gets detailed information about the usage of the specified license.

Creating a Request

data GetLicenseUsage

See: newGetLicenseUsage smart constructor.

Constructors

GetLicenseUsage'

Fields

licenseArn :: Text

Amazon Resource Name (ARN) of the license.

newGetLicenseUsage

Arguments

:: Text
$sel:licenseArn:GetLicenseUsage'

-> GetLicenseUsage

Create a value of GetLicenseUsage with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:licenseArn:GetLicenseUsage', getLicenseUsage_licenseArn - Amazon Resource Name (ARN) of the license.

Request Lenses

getLicenseUsage_licenseArn :: Lens' GetLicenseUsage Text

Amazon Resource Name (ARN) of the license.

Destructuring the Response

data GetLicenseUsageResponse

See: newGetLicenseUsageResponse smart constructor.

Constructors

GetLicenseUsageResponse'

Fields

licenseUsage :: Maybe LicenseUsage

License usage details.

httpStatus :: Int

The response's http status code.

newGetLicenseUsageResponse

Arguments

:: Int
$sel:httpStatus:GetLicenseUsageResponse'

-> GetLicenseUsageResponse

Create a value of GetLicenseUsageResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:licenseUsage:GetLicenseUsageResponse', getLicenseUsageResponse_licenseUsage - License usage details.

$sel:httpStatus:GetLicenseUsageResponse', getLicenseUsageResponse_httpStatus - The response's http status code.

Response Lenses

getLicenseUsageResponse_licenseUsage :: Lens' GetLicenseUsageResponse (Maybe LicenseUsage)

License usage details.

getLicenseUsageResponse_httpStatus :: Lens' GetLicenseUsageResponse Int

The response's http status code.