Amazonka.LicenseManager.GetServiceSettings

Description

Gets the License Manager settings for the current Region.

Creating a Request

data GetServiceSettings

See: newGetServiceSettings smart constructor.

Constructors

GetServiceSettings'

newGetServiceSettings :: GetServiceSettings

Create a value of GetServiceSettings with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

Destructuring the Response

data GetServiceSettingsResponse

See: newGetServiceSettingsResponse smart constructor.

Constructors

GetServiceSettingsResponse'

Fields

enableCrossAccountsDiscovery :: Maybe Bool

Indicates whether cross-account discovery is enabled.

snsTopicArn :: Maybe Text

SNS topic configured to receive notifications from License Manager.

licenseManagerResourceShareArn :: Maybe Text

Amazon Resource Name (ARN) of the resource share. The License Manager management account provides member accounts with access to this share.

s3BucketArn :: Maybe Text

Regional S3 bucket path for storing reports, license trail event data, discovery data, and so on.

organizationConfiguration :: Maybe OrganizationConfiguration

Indicates whether Organizations is integrated with License Manager for cross-account discovery.

httpStatus :: Int

The response's http status code.

newGetServiceSettingsResponse

Arguments

:: Int
$sel:httpStatus:GetServiceSettingsResponse'

-> GetServiceSettingsResponse

Create a value of GetServiceSettingsResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:enableCrossAccountsDiscovery:GetServiceSettingsResponse', getServiceSettingsResponse_enableCrossAccountsDiscovery - Indicates whether cross-account discovery is enabled.

$sel:snsTopicArn:GetServiceSettingsResponse', getServiceSettingsResponse_snsTopicArn - SNS topic configured to receive notifications from License Manager.

$sel:licenseManagerResourceShareArn:GetServiceSettingsResponse', getServiceSettingsResponse_licenseManagerResourceShareArn - Amazon Resource Name (ARN) of the resource share. The License Manager management account provides member accounts with access to this share.

$sel:s3BucketArn:GetServiceSettingsResponse', getServiceSettingsResponse_s3BucketArn - Regional S3 bucket path for storing reports, license trail event data, discovery data, and so on.

$sel:organizationConfiguration:GetServiceSettingsResponse', getServiceSettingsResponse_organizationConfiguration - Indicates whether Organizations is integrated with License Manager for cross-account discovery.

$sel:httpStatus:GetServiceSettingsResponse', getServiceSettingsResponse_httpStatus - The response's http status code.

Response Lenses

getServiceSettingsResponse_enableCrossAccountsDiscovery :: Lens' GetServiceSettingsResponse (Maybe Bool)

Indicates whether cross-account discovery is enabled.

getServiceSettingsResponse_snsTopicArn :: Lens' GetServiceSettingsResponse (Maybe Text)

SNS topic configured to receive notifications from License Manager.

getServiceSettingsResponse_licenseManagerResourceShareArn :: Lens' GetServiceSettingsResponse (Maybe Text)

Amazon Resource Name (ARN) of the resource share. The License Manager management account provides member accounts with access to this share.

getServiceSettingsResponse_s3BucketArn :: Lens' GetServiceSettingsResponse (Maybe Text)

Regional S3 bucket path for storing reports, license trail event data, discovery data, and so on.

getServiceSettingsResponse_organizationConfiguration :: Lens' GetServiceSettingsResponse (Maybe OrganizationConfiguration)

Indicates whether Organizations is integrated with License Manager for cross-account discovery.

getServiceSettingsResponse_httpStatus :: Lens' GetServiceSettingsResponse Int

The response's http status code.