Amazonka.LicenseManager.ListAssociationsForLicenseConfiguration

Description

Lists the resource associations for the specified license configuration.

Resource associations need not consume licenses from a license configuration. For example, an AMI or a stopped instance might not consume a license (depending on the license rules).

This operation returns paginated results.

Creating a Request

data ListAssociationsForLicenseConfiguration

See: newListAssociationsForLicenseConfiguration smart constructor.

Constructors

ListAssociationsForLicenseConfiguration'

Fields

nextToken :: Maybe Text

Token for the next set of results.

maxResults :: Maybe Int

Maximum number of results to return in a single call.

licenseConfigurationArn :: Text

Amazon Resource Name (ARN) of a license configuration.

newListAssociationsForLicenseConfiguration

Arguments

:: Text
$sel:licenseConfigurationArn:ListAssociationsForLicenseConfiguration'

-> ListAssociationsForLicenseConfiguration

Create a value of ListAssociationsForLicenseConfiguration with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:nextToken:ListAssociationsForLicenseConfiguration', listAssociationsForLicenseConfiguration_nextToken - Token for the next set of results.

$sel:maxResults:ListAssociationsForLicenseConfiguration', listAssociationsForLicenseConfiguration_maxResults - Maximum number of results to return in a single call.

$sel:licenseConfigurationArn:ListAssociationsForLicenseConfiguration', listAssociationsForLicenseConfiguration_licenseConfigurationArn - Amazon Resource Name (ARN) of a license configuration.

Request Lenses

listAssociationsForLicenseConfiguration_nextToken :: Lens' ListAssociationsForLicenseConfiguration (Maybe Text)

Token for the next set of results.

listAssociationsForLicenseConfiguration_maxResults :: Lens' ListAssociationsForLicenseConfiguration (Maybe Int)

Maximum number of results to return in a single call.

listAssociationsForLicenseConfiguration_licenseConfigurationArn :: Lens' ListAssociationsForLicenseConfiguration Text

Amazon Resource Name (ARN) of a license configuration.

Destructuring the Response

data ListAssociationsForLicenseConfigurationResponse

See: newListAssociationsForLicenseConfigurationResponse smart constructor.

Constructors

ListAssociationsForLicenseConfigurationResponse'

Fields

licenseConfigurationAssociations :: Maybe [LicenseConfigurationAssociation]

Information about the associations for the license configuration.

nextToken :: Maybe Text

Token for the next set of results.

httpStatus :: Int

The response's http status code.

newListAssociationsForLicenseConfigurationResponse

Arguments

:: Int
$sel:httpStatus:ListAssociationsForLicenseConfigurationResponse'

-> ListAssociationsForLicenseConfigurationResponse

Create a value of ListAssociationsForLicenseConfigurationResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:licenseConfigurationAssociations:ListAssociationsForLicenseConfigurationResponse', listAssociationsForLicenseConfigurationResponse_licenseConfigurationAssociations - Information about the associations for the license configuration.

$sel:nextToken:ListAssociationsForLicenseConfiguration', listAssociationsForLicenseConfigurationResponse_nextToken - Token for the next set of results.

$sel:httpStatus:ListAssociationsForLicenseConfigurationResponse', listAssociationsForLicenseConfigurationResponse_httpStatus - The response's http status code.

Response Lenses

listAssociationsForLicenseConfigurationResponse_licenseConfigurationAssociations :: Lens' ListAssociationsForLicenseConfigurationResponse (Maybe [LicenseConfigurationAssociation])

Information about the associations for the license configuration.

listAssociationsForLicenseConfigurationResponse_nextToken :: Lens' ListAssociationsForLicenseConfigurationResponse (Maybe Text)

Token for the next set of results.

listAssociationsForLicenseConfigurationResponse_httpStatus :: Lens' ListAssociationsForLicenseConfigurationResponse Int

The response's http status code.