Amazonka.LicenseManager.Types.Entitlement

Description

Documentation

data Entitlement

Describes a resource entitled for use with a license.

See: newEntitlement smart constructor.

Constructors

Entitlement'

Fields

maxCount :: Maybe Integer

Maximum entitlement count. Use if the unit is not None.

value :: Maybe Text

Entitlement resource. Use only if the unit is None.

overage :: Maybe Bool

Indicates whether overages are allowed.

allowCheckIn :: Maybe Bool

Indicates whether check-ins are allowed.

name :: Text

Entitlement name.

unit :: EntitlementUnit

Entitlement unit.

newEntitlement

Arguments

:: Text
$sel:name:Entitlement'

-> EntitlementUnit
$sel:unit:Entitlement'

-> Entitlement

Create a value of Entitlement with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:maxCount:Entitlement', entitlement_maxCount - Maximum entitlement count. Use if the unit is not None.

$sel:value:Entitlement', entitlement_value - Entitlement resource. Use only if the unit is None.

$sel:overage:Entitlement', entitlement_overage - Indicates whether overages are allowed.

$sel:allowCheckIn:Entitlement', entitlement_allowCheckIn - Indicates whether check-ins are allowed.

$sel:name:Entitlement', entitlement_name - Entitlement name.

$sel:unit:Entitlement', entitlement_unit - Entitlement unit.

entitlement_maxCount :: Lens' Entitlement (Maybe Integer)

Maximum entitlement count. Use if the unit is not None.

entitlement_value :: Lens' Entitlement (Maybe Text)

Entitlement resource. Use only if the unit is None.

entitlement_overage :: Lens' Entitlement (Maybe Bool)

Indicates whether overages are allowed.

entitlement_allowCheckIn :: Lens' Entitlement (Maybe Bool)

Indicates whether check-ins are allowed.

entitlement_name :: Lens' Entitlement Text

Entitlement name.

entitlement_unit :: Lens' Entitlement EntitlementUnit

Entitlement unit.