Amazonka.LicenseManager.Types.Grant

Description

Documentation

data Grant

Describes a grant.

See: newGrant smart constructor.

Constructors

Grant'

Fields

statusReason :: Maybe Text

Grant status reason.

grantArn :: Text

Amazon Resource Name (ARN) of the grant.

grantName :: Text

Grant name.

parentArn :: Text

Parent ARN.

licenseArn :: Text

License ARN.

granteePrincipalArn :: Text

The grantee principal ARN.

homeRegion :: Text

Home Region of the grant.

grantStatus :: GrantStatus

Grant status.

version :: Text

Grant version.

grantedOperations :: NonEmpty AllowedOperation

Granted operations.

newGrant

Arguments

:: Text
$sel:grantArn:Grant'

-> Text
$sel:grantName:Grant'

-> Text
$sel:parentArn:Grant'

-> Text
$sel:licenseArn:Grant'

-> Text
$sel:granteePrincipalArn:Grant'

-> Text
$sel:homeRegion:Grant'

-> GrantStatus
$sel:grantStatus:Grant'

-> Text
$sel:version:Grant'

-> NonEmpty AllowedOperation
$sel:grantedOperations:Grant'

-> Grant

Create a value of Grant with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:statusReason:Grant', grant_statusReason - Grant status reason.

$sel:grantArn:Grant', grant_grantArn - Amazon Resource Name (ARN) of the grant.

$sel:grantName:Grant', grant_grantName - Grant name.

$sel:parentArn:Grant', grant_parentArn - Parent ARN.

$sel:licenseArn:Grant', grant_licenseArn - License ARN.

$sel:granteePrincipalArn:Grant', grant_granteePrincipalArn - The grantee principal ARN.

$sel:homeRegion:Grant', grant_homeRegion - Home Region of the grant.

$sel:grantStatus:Grant', grant_grantStatus - Grant status.

$sel:version:Grant', grant_version - Grant version.

$sel:grantedOperations:Grant', grant_grantedOperations - Granted operations.

grant_statusReason :: Lens' Grant (Maybe Text)

Grant status reason.

grant_grantArn :: Lens' Grant Text

Amazon Resource Name (ARN) of the grant.

grant_grantName :: Lens' Grant Text

Grant name.

grant_parentArn :: Lens' Grant Text

Parent ARN.

grant_licenseArn :: Lens' Grant Text

License ARN.

grant_granteePrincipalArn :: Lens' Grant Text

The grantee principal ARN.

grant_homeRegion :: Lens' Grant Text

Home Region of the grant.

grant_grantStatus :: Lens' Grant GrantStatus

Grant status.

grant_version :: Lens' Grant Text

Grant version.

grant_grantedOperations :: Lens' Grant (NonEmpty AllowedOperation)

Granted operations.