Amazonka.LicenseManager.Types.GrantedLicense

Description

Documentation

data GrantedLicense

Describes a license that is granted to a grantee.

See: newGrantedLicense smart constructor.

Constructors

GrantedLicense'

Fields

status :: Maybe LicenseStatus

Granted license status.

productName :: Maybe Text

Product name.

licenseName :: Maybe Text

License name.

productSKU :: Maybe Text

Product SKU.

homeRegion :: Maybe Text

Home Region of the granted license.

receivedMetadata :: Maybe ReceivedMetadata

Granted license received metadata.

version :: Maybe Text

Version of the granted license.

entitlements :: Maybe [Entitlement]

License entitlements.

validity :: Maybe DatetimeRange

Date and time range during which the granted license is valid, in ISO8601-UTC format.

consumptionConfiguration :: Maybe ConsumptionConfiguration

Configuration for consumption of the license.

beneficiary :: Maybe Text

Granted license beneficiary.

issuer :: Maybe IssuerDetails

Granted license issuer.

licenseArn :: Maybe Text

Amazon Resource Name (ARN) of the license.

licenseMetadata :: Maybe [Metadata]

Granted license metadata.

createTime :: Maybe Text

Creation time of the granted license.

newGrantedLicense :: GrantedLicense

Create a value of GrantedLicense with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:status:GrantedLicense', grantedLicense_status - Granted license status.

$sel:productName:GrantedLicense', grantedLicense_productName - Product name.

$sel:licenseName:GrantedLicense', grantedLicense_licenseName - License name.

$sel:productSKU:GrantedLicense', grantedLicense_productSKU - Product SKU.

$sel:homeRegion:GrantedLicense', grantedLicense_homeRegion - Home Region of the granted license.

$sel:receivedMetadata:GrantedLicense', grantedLicense_receivedMetadata - Granted license received metadata.

$sel:version:GrantedLicense', grantedLicense_version - Version of the granted license.

$sel:entitlements:GrantedLicense', grantedLicense_entitlements - License entitlements.

$sel:validity:GrantedLicense', grantedLicense_validity - Date and time range during which the granted license is valid, in ISO8601-UTC format.

$sel:consumptionConfiguration:GrantedLicense', grantedLicense_consumptionConfiguration - Configuration for consumption of the license.

$sel:beneficiary:GrantedLicense', grantedLicense_beneficiary - Granted license beneficiary.

$sel:issuer:GrantedLicense', grantedLicense_issuer - Granted license issuer.

$sel:licenseArn:GrantedLicense', grantedLicense_licenseArn - Amazon Resource Name (ARN) of the license.

$sel:licenseMetadata:GrantedLicense', grantedLicense_licenseMetadata - Granted license metadata.

$sel:createTime:GrantedLicense', grantedLicense_createTime - Creation time of the granted license.

grantedLicense_status :: Lens' GrantedLicense (Maybe LicenseStatus)

Granted license status.

grantedLicense_productName :: Lens' GrantedLicense (Maybe Text)

Product name.

grantedLicense_licenseName :: Lens' GrantedLicense (Maybe Text)

License name.

grantedLicense_productSKU :: Lens' GrantedLicense (Maybe Text)

Product SKU.

grantedLicense_homeRegion :: Lens' GrantedLicense (Maybe Text)

Home Region of the granted license.

grantedLicense_receivedMetadata :: Lens' GrantedLicense (Maybe ReceivedMetadata)

Granted license received metadata.

grantedLicense_version :: Lens' GrantedLicense (Maybe Text)

Version of the granted license.

grantedLicense_entitlements :: Lens' GrantedLicense (Maybe [Entitlement])

License entitlements.

grantedLicense_validity :: Lens' GrantedLicense (Maybe DatetimeRange)

Date and time range during which the granted license is valid, in ISO8601-UTC format.

grantedLicense_consumptionConfiguration :: Lens' GrantedLicense (Maybe ConsumptionConfiguration)

Configuration for consumption of the license.

grantedLicense_beneficiary :: Lens' GrantedLicense (Maybe Text)

Granted license beneficiary.

grantedLicense_issuer :: Lens' GrantedLicense (Maybe IssuerDetails)

Granted license issuer.

grantedLicense_licenseArn :: Lens' GrantedLicense (Maybe Text)

Amazon Resource Name (ARN) of the license.

grantedLicense_licenseMetadata :: Lens' GrantedLicense (Maybe [Metadata])

Granted license metadata.

grantedLicense_createTime :: Lens' GrantedLicense (Maybe Text)

Creation time of the granted license.