Amazonka.LicenseManager.Types.IssuerDetails

Description

Documentation

data IssuerDetails

Details associated with the issuer of a license.

See: newIssuerDetails smart constructor.

Constructors

IssuerDetails'

Fields

keyFingerprint :: Maybe Text

Issuer key fingerprint.

signKey :: Maybe Text

Asymmetric KMS key from Key Management Service. The KMS key must have a key usage of sign and verify, and support the RSASSA-PSS SHA-256 signing algorithm.

name :: Maybe Text

Issuer name.

newIssuerDetails :: IssuerDetails

Create a value of IssuerDetails with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:keyFingerprint:IssuerDetails', issuerDetails_keyFingerprint - Issuer key fingerprint.

$sel:signKey:IssuerDetails', issuerDetails_signKey - Asymmetric KMS key from Key Management Service. The KMS key must have a key usage of sign and verify, and support the RSASSA-PSS SHA-256 signing algorithm.

$sel:name:IssuerDetails', issuerDetails_name - Issuer name.

issuerDetails_keyFingerprint :: Lens' IssuerDetails (Maybe Text)

Issuer key fingerprint.

issuerDetails_signKey :: Lens' IssuerDetails (Maybe Text)

Asymmetric KMS key from Key Management Service. The KMS key must have a key usage of sign and verify, and support the RSASSA-PSS SHA-256 signing algorithm.

issuerDetails_name :: Lens' IssuerDetails (Maybe Text)

Issuer name.