Amazonka.LicenseManager.Types.LicenseConfiguration

Description

Documentation

data LicenseConfiguration

A license configuration is an abstraction of a customer license agreement that can be consumed and enforced by License Manager. Components include specifications for the license type (licensing by instance, socket, CPU, or vCPU), allowed tenancy (shared tenancy, Dedicated Instance, Dedicated Host, or all of these), host affinity (how long a VM must be associated with a host), and the number of licenses purchased and used.

See: newLicenseConfiguration smart constructor.

Constructors

LicenseConfiguration'

Fields

status :: Maybe Text

Status of the license configuration.

ownerAccountId :: Maybe Text

Account ID of the license configuration's owner.

consumedLicenseSummaryList :: Maybe [ConsumedLicenseSummary]

Summaries for licenses consumed by various resources.

licenseCount :: Maybe Integer

Number of licenses managed by the license configuration.

managedResourceSummaryList :: Maybe [ManagedResourceSummary]

Summaries for managed resources.

name :: Maybe Text

Name of the license configuration.

licenseCountHardLimit :: Maybe Bool

Number of available licenses as a hard limit.

disassociateWhenNotFound :: Maybe Bool

When true, disassociates a resource when software is uninstalled.

productInformationList :: Maybe [ProductInformation]

Product information.

licenseCountingType :: Maybe LicenseCountingType

Dimension to use to track the license inventory.

automatedDiscoveryInformation :: Maybe AutomatedDiscoveryInformation

Automated discovery information.

consumedLicenses :: Maybe Integer

Number of licenses consumed.

licenseRules :: Maybe [Text]

License rules.

licenseConfigurationId :: Maybe Text

Unique ID of the license configuration.

description :: Maybe Text

Description of the license configuration.

licenseConfigurationArn :: Maybe Text

Amazon Resource Name (ARN) of the license configuration.

newLicenseConfiguration :: LicenseConfiguration

Create a value of LicenseConfiguration with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:status:LicenseConfiguration', licenseConfiguration_status - Status of the license configuration.

$sel:ownerAccountId:LicenseConfiguration', licenseConfiguration_ownerAccountId - Account ID of the license configuration's owner.

$sel:consumedLicenseSummaryList:LicenseConfiguration', licenseConfiguration_consumedLicenseSummaryList - Summaries for licenses consumed by various resources.

$sel:licenseCount:LicenseConfiguration', licenseConfiguration_licenseCount - Number of licenses managed by the license configuration.

$sel:managedResourceSummaryList:LicenseConfiguration', licenseConfiguration_managedResourceSummaryList - Summaries for managed resources.

$sel:name:LicenseConfiguration', licenseConfiguration_name - Name of the license configuration.

$sel:licenseCountHardLimit:LicenseConfiguration', licenseConfiguration_licenseCountHardLimit - Number of available licenses as a hard limit.

$sel:disassociateWhenNotFound:LicenseConfiguration', licenseConfiguration_disassociateWhenNotFound - When true, disassociates a resource when software is uninstalled.

$sel:productInformationList:LicenseConfiguration', licenseConfiguration_productInformationList - Product information.

$sel:licenseCountingType:LicenseConfiguration', licenseConfiguration_licenseCountingType - Dimension to use to track the license inventory.

$sel:automatedDiscoveryInformation:LicenseConfiguration', licenseConfiguration_automatedDiscoveryInformation - Automated discovery information.

$sel:consumedLicenses:LicenseConfiguration', licenseConfiguration_consumedLicenses - Number of licenses consumed.

$sel:licenseRules:LicenseConfiguration', licenseConfiguration_licenseRules - License rules.

$sel:licenseConfigurationId:LicenseConfiguration', licenseConfiguration_licenseConfigurationId - Unique ID of the license configuration.

$sel:description:LicenseConfiguration', licenseConfiguration_description - Description of the license configuration.

$sel:licenseConfigurationArn:LicenseConfiguration', licenseConfiguration_licenseConfigurationArn - Amazon Resource Name (ARN) of the license configuration.

licenseConfiguration_status :: Lens' LicenseConfiguration (Maybe Text)

Status of the license configuration.

licenseConfiguration_ownerAccountId :: Lens' LicenseConfiguration (Maybe Text)

Account ID of the license configuration's owner.

licenseConfiguration_consumedLicenseSummaryList :: Lens' LicenseConfiguration (Maybe [ConsumedLicenseSummary])

Summaries for licenses consumed by various resources.

licenseConfiguration_licenseCount :: Lens' LicenseConfiguration (Maybe Integer)

Number of licenses managed by the license configuration.

licenseConfiguration_managedResourceSummaryList :: Lens' LicenseConfiguration (Maybe [ManagedResourceSummary])

Summaries for managed resources.

licenseConfiguration_name :: Lens' LicenseConfiguration (Maybe Text)

Name of the license configuration.

licenseConfiguration_licenseCountHardLimit :: Lens' LicenseConfiguration (Maybe Bool)

Number of available licenses as a hard limit.

licenseConfiguration_disassociateWhenNotFound :: Lens' LicenseConfiguration (Maybe Bool)

When true, disassociates a resource when software is uninstalled.

licenseConfiguration_productInformationList :: Lens' LicenseConfiguration (Maybe [ProductInformation])

Product information.

licenseConfiguration_licenseCountingType :: Lens' LicenseConfiguration (Maybe LicenseCountingType)

Dimension to use to track the license inventory.

licenseConfiguration_automatedDiscoveryInformation :: Lens' LicenseConfiguration (Maybe AutomatedDiscoveryInformation)

Automated discovery information.

licenseConfiguration_consumedLicenses :: Lens' LicenseConfiguration (Maybe Integer)

Number of licenses consumed.

licenseConfiguration_licenseRules :: Lens' LicenseConfiguration (Maybe [Text])

License rules.

licenseConfiguration_licenseConfigurationId :: Lens' LicenseConfiguration (Maybe Text)

Unique ID of the license configuration.

licenseConfiguration_description :: Lens' LicenseConfiguration (Maybe Text)

Description of the license configuration.

licenseConfiguration_licenseConfigurationArn :: Lens' LicenseConfiguration (Maybe Text)

Amazon Resource Name (ARN) of the license configuration.