Amazonka.LicenseManager.Types.LicenseConfigurationAssociation

Description

Documentation

data LicenseConfigurationAssociation

Describes an association with a license configuration.

See: newLicenseConfigurationAssociation smart constructor.

Constructors

LicenseConfigurationAssociation'

Fields

resourceType :: Maybe ResourceType

Type of server resource.

amiAssociationScope :: Maybe Text

Scope of AMI associations. The possible value is cross-account.

associationTime :: Maybe POSIX

Time when the license configuration was associated with the resource.

resourceArn :: Maybe Text

Amazon Resource Name (ARN) of the resource.

resourceOwnerId :: Maybe Text

ID of the Amazon Web Services account that owns the resource consuming licenses.

newLicenseConfigurationAssociation :: LicenseConfigurationAssociation

Create a value of LicenseConfigurationAssociation with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:resourceType:LicenseConfigurationAssociation', licenseConfigurationAssociation_resourceType - Type of server resource.

$sel:amiAssociationScope:LicenseConfigurationAssociation', licenseConfigurationAssociation_amiAssociationScope - Scope of AMI associations. The possible value is cross-account.

$sel:associationTime:LicenseConfigurationAssociation', licenseConfigurationAssociation_associationTime - Time when the license configuration was associated with the resource.

$sel:resourceArn:LicenseConfigurationAssociation', licenseConfigurationAssociation_resourceArn - Amazon Resource Name (ARN) of the resource.

$sel:resourceOwnerId:LicenseConfigurationAssociation', licenseConfigurationAssociation_resourceOwnerId - ID of the Amazon Web Services account that owns the resource consuming licenses.

licenseConfigurationAssociation_resourceType :: Lens' LicenseConfigurationAssociation (Maybe ResourceType)

Type of server resource.

licenseConfigurationAssociation_amiAssociationScope :: Lens' LicenseConfigurationAssociation (Maybe Text)

Scope of AMI associations. The possible value is cross-account.

licenseConfigurationAssociation_associationTime :: Lens' LicenseConfigurationAssociation (Maybe UTCTime)

Time when the license configuration was associated with the resource.

licenseConfigurationAssociation_resourceArn :: Lens' LicenseConfigurationAssociation (Maybe Text)

Amazon Resource Name (ARN) of the resource.

licenseConfigurationAssociation_resourceOwnerId :: Lens' LicenseConfigurationAssociation (Maybe Text)

ID of the Amazon Web Services account that owns the resource consuming licenses.