Amazonka.LicenseManager.Types.LicenseConversionTask

Description

Documentation

data LicenseConversionTask

Information about a license type conversion task.

See: newLicenseConversionTask smart constructor.

Constructors

LicenseConversionTask'

Fields

status :: Maybe LicenseConversionTaskStatus

The status of the conversion task.

startTime :: Maybe POSIX

The time the conversion task was started at.

destinationLicenseContext :: Maybe LicenseConversionContext

Information about the license type this conversion task converted to.

licenseConversionTaskId :: Maybe Text

The ID of the license type conversion task.

resourceArn :: Maybe Text

The Amazon Resource Name (ARN) of the resource associated with the license type conversion task.

statusMessage :: Maybe Text

The status message for the conversion task.

endTime :: Maybe POSIX

The time the conversion task was completed.

licenseConversionTime :: Maybe POSIX

The time the usage operation value of the resource was changed.

sourceLicenseContext :: Maybe LicenseConversionContext

Information about the license type this conversion task converted from.

newLicenseConversionTask :: LicenseConversionTask

Create a value of LicenseConversionTask with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:status:LicenseConversionTask', licenseConversionTask_status - The status of the conversion task.

$sel:startTime:LicenseConversionTask', licenseConversionTask_startTime - The time the conversion task was started at.

$sel:destinationLicenseContext:LicenseConversionTask', licenseConversionTask_destinationLicenseContext - Information about the license type this conversion task converted to.

$sel:licenseConversionTaskId:LicenseConversionTask', licenseConversionTask_licenseConversionTaskId - The ID of the license type conversion task.

$sel:resourceArn:LicenseConversionTask', licenseConversionTask_resourceArn - The Amazon Resource Name (ARN) of the resource associated with the license type conversion task.

$sel:statusMessage:LicenseConversionTask', licenseConversionTask_statusMessage - The status message for the conversion task.

$sel:endTime:LicenseConversionTask', licenseConversionTask_endTime - The time the conversion task was completed.

$sel:licenseConversionTime:LicenseConversionTask', licenseConversionTask_licenseConversionTime - The time the usage operation value of the resource was changed.

$sel:sourceLicenseContext:LicenseConversionTask', licenseConversionTask_sourceLicenseContext - Information about the license type this conversion task converted from.

licenseConversionTask_status :: Lens' LicenseConversionTask (Maybe LicenseConversionTaskStatus)

The status of the conversion task.

licenseConversionTask_startTime :: Lens' LicenseConversionTask (Maybe UTCTime)

The time the conversion task was started at.

licenseConversionTask_destinationLicenseContext :: Lens' LicenseConversionTask (Maybe LicenseConversionContext)

Information about the license type this conversion task converted to.

licenseConversionTask_licenseConversionTaskId :: Lens' LicenseConversionTask (Maybe Text)

The ID of the license type conversion task.

licenseConversionTask_resourceArn :: Lens' LicenseConversionTask (Maybe Text)

The Amazon Resource Name (ARN) of the resource associated with the license type conversion task.

licenseConversionTask_statusMessage :: Lens' LicenseConversionTask (Maybe Text)

The status message for the conversion task.

licenseConversionTask_endTime :: Lens' LicenseConversionTask (Maybe UTCTime)

The time the conversion task was completed.

licenseConversionTask_licenseConversionTime :: Lens' LicenseConversionTask (Maybe UTCTime)

The time the usage operation value of the resource was changed.

licenseConversionTask_sourceLicenseContext :: Lens' LicenseConversionTask (Maybe LicenseConversionContext)

Information about the license type this conversion task converted from.