Amazonka.LicenseManager.Types.LicenseOperationFailure

Description

Documentation

data LicenseOperationFailure

Describes the failure of a license operation.

See: newLicenseOperationFailure smart constructor.

Constructors

LicenseOperationFailure'

Fields

resourceType :: Maybe ResourceType

Resource type.

operationRequestedBy :: Maybe Text

The requester is "License Manager Automated Discovery".

resourceArn :: Maybe Text

Amazon Resource Name (ARN) of the resource.

metadataList :: Maybe [Metadata]

Reserved.

operationName :: Maybe Text

Name of the operation.

failureTime :: Maybe POSIX

Failure time.

errorMessage :: Maybe Text

Error message.

resourceOwnerId :: Maybe Text

ID of the Amazon Web Services account that owns the resource.

newLicenseOperationFailure :: LicenseOperationFailure

Create a value of LicenseOperationFailure with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:resourceType:LicenseOperationFailure', licenseOperationFailure_resourceType - Resource type.

$sel:operationRequestedBy:LicenseOperationFailure', licenseOperationFailure_operationRequestedBy - The requester is "License Manager Automated Discovery".

$sel:resourceArn:LicenseOperationFailure', licenseOperationFailure_resourceArn - Amazon Resource Name (ARN) of the resource.

$sel:metadataList:LicenseOperationFailure', licenseOperationFailure_metadataList - Reserved.

$sel:operationName:LicenseOperationFailure', licenseOperationFailure_operationName - Name of the operation.

$sel:failureTime:LicenseOperationFailure', licenseOperationFailure_failureTime - Failure time.

$sel:errorMessage:LicenseOperationFailure', licenseOperationFailure_errorMessage - Error message.

$sel:resourceOwnerId:LicenseOperationFailure', licenseOperationFailure_resourceOwnerId - ID of the Amazon Web Services account that owns the resource.

licenseOperationFailure_resourceType :: Lens' LicenseOperationFailure (Maybe ResourceType)

Resource type.

licenseOperationFailure_operationRequestedBy :: Lens' LicenseOperationFailure (Maybe Text)

The requester is "License Manager Automated Discovery".

licenseOperationFailure_resourceArn :: Lens' LicenseOperationFailure (Maybe Text)

Amazon Resource Name (ARN) of the resource.

licenseOperationFailure_metadataList :: Lens' LicenseOperationFailure (Maybe [Metadata])

Reserved.

licenseOperationFailure_operationName :: Lens' LicenseOperationFailure (Maybe Text)

Name of the operation.

licenseOperationFailure_failureTime :: Lens' LicenseOperationFailure (Maybe UTCTime)

Failure time.

licenseOperationFailure_errorMessage :: Lens' LicenseOperationFailure (Maybe Text)

Error message.

licenseOperationFailure_resourceOwnerId :: Lens' LicenseOperationFailure (Maybe Text)

ID of the Amazon Web Services account that owns the resource.