Amazonka.LicenseManager.UpdateLicenseConfiguration

Description

Modifies the attributes of an existing license configuration.

Creating a Request

data UpdateLicenseConfiguration

See: newUpdateLicenseConfiguration smart constructor.

Constructors

UpdateLicenseConfiguration'

Fields

licenseCount :: Maybe Integer

New number of licenses managed by the license configuration.

name :: Maybe Text

New name of the license configuration.

licenseConfigurationStatus :: Maybe LicenseConfigurationStatus

New status of the license configuration.

licenseCountHardLimit :: Maybe Bool

New hard limit of the number of available licenses.

disassociateWhenNotFound :: Maybe Bool

When true, disassociates a resource when software is uninstalled.

productInformationList :: Maybe [ProductInformation]

New product information.

licenseRules :: Maybe [Text]

New license rule. The only rule that you can add after you create a license configuration is licenseAffinityToHost.

description :: Maybe Text

New description of the license configuration.

licenseConfigurationArn :: Text

Amazon Resource Name (ARN) of the license configuration.

newUpdateLicenseConfiguration

Arguments

:: Text
$sel:licenseConfigurationArn:UpdateLicenseConfiguration'

-> UpdateLicenseConfiguration

Create a value of UpdateLicenseConfiguration with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:licenseCount:UpdateLicenseConfiguration', updateLicenseConfiguration_licenseCount - New number of licenses managed by the license configuration.

$sel:name:UpdateLicenseConfiguration', updateLicenseConfiguration_name - New name of the license configuration.

$sel:licenseConfigurationStatus:UpdateLicenseConfiguration', updateLicenseConfiguration_licenseConfigurationStatus - New status of the license configuration.

$sel:licenseCountHardLimit:UpdateLicenseConfiguration', updateLicenseConfiguration_licenseCountHardLimit - New hard limit of the number of available licenses.

$sel:disassociateWhenNotFound:UpdateLicenseConfiguration', updateLicenseConfiguration_disassociateWhenNotFound - When true, disassociates a resource when software is uninstalled.

$sel:productInformationList:UpdateLicenseConfiguration', updateLicenseConfiguration_productInformationList - New product information.

$sel:licenseRules:UpdateLicenseConfiguration', updateLicenseConfiguration_licenseRules - New license rule. The only rule that you can add after you create a license configuration is licenseAffinityToHost.

$sel:description:UpdateLicenseConfiguration', updateLicenseConfiguration_description - New description of the license configuration.

$sel:licenseConfigurationArn:UpdateLicenseConfiguration', updateLicenseConfiguration_licenseConfigurationArn - Amazon Resource Name (ARN) of the license configuration.

Request Lenses

updateLicenseConfiguration_licenseCount :: Lens' UpdateLicenseConfiguration (Maybe Integer)

New number of licenses managed by the license configuration.

updateLicenseConfiguration_name :: Lens' UpdateLicenseConfiguration (Maybe Text)

New name of the license configuration.

updateLicenseConfiguration_licenseConfigurationStatus :: Lens' UpdateLicenseConfiguration (Maybe LicenseConfigurationStatus)

New status of the license configuration.

updateLicenseConfiguration_licenseCountHardLimit :: Lens' UpdateLicenseConfiguration (Maybe Bool)

New hard limit of the number of available licenses.

updateLicenseConfiguration_disassociateWhenNotFound :: Lens' UpdateLicenseConfiguration (Maybe Bool)

When true, disassociates a resource when software is uninstalled.

updateLicenseConfiguration_productInformationList :: Lens' UpdateLicenseConfiguration (Maybe [ProductInformation])

New product information.

updateLicenseConfiguration_licenseRules :: Lens' UpdateLicenseConfiguration (Maybe [Text])

New license rule. The only rule that you can add after you create a license configuration is licenseAffinityToHost.

updateLicenseConfiguration_description :: Lens' UpdateLicenseConfiguration (Maybe Text)

New description of the license configuration.

updateLicenseConfiguration_licenseConfigurationArn :: Lens' UpdateLicenseConfiguration Text

Amazon Resource Name (ARN) of the license configuration.

Destructuring the Response

data UpdateLicenseConfigurationResponse

See: newUpdateLicenseConfigurationResponse smart constructor.

Constructors

UpdateLicenseConfigurationResponse'

Fields

httpStatus :: Int

The response's http status code.

newUpdateLicenseConfigurationResponse

Arguments

:: Int
$sel:httpStatus:UpdateLicenseConfigurationResponse'

-> UpdateLicenseConfigurationResponse

Create a value of UpdateLicenseConfigurationResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:httpStatus:UpdateLicenseConfigurationResponse', updateLicenseConfigurationResponse_httpStatus - The response's http status code.

Response Lenses

updateLicenseConfigurationResponse_httpStatus :: Lens' UpdateLicenseConfigurationResponse Int

The response's http status code.