Amazonka.Lightsail.AttachLoadBalancerTlsCertificate

Description

Attaches a Transport Layer Security (TLS) certificate to your load balancer. TLS is just an updated, more secure version of Secure Socket Layer (SSL).

Once you create and validate your certificate, you can attach it to your load balancer. You can also use this API to rotate the certificates on your account. Use the AttachLoadBalancerTlsCertificate action with the non-attached certificate, and it will replace the existing one and become the attached certificate.

The AttachLoadBalancerTlsCertificate operation supports tag-based access control via resource tags applied to the resource identified by load balancer name. For more information, see the Amazon Lightsail Developer Guide.

Creating a Request

data AttachLoadBalancerTlsCertificate

See: newAttachLoadBalancerTlsCertificate smart constructor.

Constructors

AttachLoadBalancerTlsCertificate'

Fields

loadBalancerName :: Text

The name of the load balancer to which you want to associate the SSL/TLS certificate.

certificateName :: Text

The name of your SSL/TLS certificate.

newAttachLoadBalancerTlsCertificate

Arguments

:: Text
$sel:loadBalancerName:AttachLoadBalancerTlsCertificate'

-> Text
$sel:certificateName:AttachLoadBalancerTlsCertificate'

-> AttachLoadBalancerTlsCertificate

Create a value of AttachLoadBalancerTlsCertificate with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:loadBalancerName:AttachLoadBalancerTlsCertificate', attachLoadBalancerTlsCertificate_loadBalancerName - The name of the load balancer to which you want to associate the SSL/TLS certificate.

$sel:certificateName:AttachLoadBalancerTlsCertificate', attachLoadBalancerTlsCertificate_certificateName - The name of your SSL/TLS certificate.

Request Lenses

attachLoadBalancerTlsCertificate_loadBalancerName :: Lens' AttachLoadBalancerTlsCertificate Text

The name of the load balancer to which you want to associate the SSL/TLS certificate.

attachLoadBalancerTlsCertificate_certificateName :: Lens' AttachLoadBalancerTlsCertificate Text

The name of your SSL/TLS certificate.

Destructuring the Response

data AttachLoadBalancerTlsCertificateResponse

See: newAttachLoadBalancerTlsCertificateResponse smart constructor.

Constructors

AttachLoadBalancerTlsCertificateResponse'

Fields

operations :: Maybe [Operation]

An array of objects that describe the result of the action, such as the status of the request, the timestamp of the request, and the resources affected by the request.

These SSL/TLS certificates are only usable by Lightsail load balancers. You can't get the certificate and use it for another purpose.

httpStatus :: Int

The response's http status code.

newAttachLoadBalancerTlsCertificateResponse

Arguments

:: Int
$sel:httpStatus:AttachLoadBalancerTlsCertificateResponse'

-> AttachLoadBalancerTlsCertificateResponse

Create a value of AttachLoadBalancerTlsCertificateResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:operations:AttachLoadBalancerTlsCertificateResponse', attachLoadBalancerTlsCertificateResponse_operations - An array of objects that describe the result of the action, such as the status of the request, the timestamp of the request, and the resources affected by the request.

These SSL/TLS certificates are only usable by Lightsail load balancers. You can't get the certificate and use it for another purpose.

$sel:httpStatus:AttachLoadBalancerTlsCertificateResponse', attachLoadBalancerTlsCertificateResponse_httpStatus - The response's http status code.

Response Lenses

attachLoadBalancerTlsCertificateResponse_operations :: Lens' AttachLoadBalancerTlsCertificateResponse (Maybe [Operation])

An array of objects that describe the result of the action, such as the status of the request, the timestamp of the request, and the resources affected by the request.

These SSL/TLS certificates are only usable by Lightsail load balancers. You can't get the certificate and use it for another purpose.

attachLoadBalancerTlsCertificateResponse_httpStatus :: Lens' AttachLoadBalancerTlsCertificateResponse Int

The response's http status code.