Amazonka.Lightsail.CopySnapshot

Description

Copies a manual snapshot of an instance or disk as another manual snapshot, or copies an automatic snapshot of an instance or disk as a manual snapshot. This operation can also be used to copy a manual or automatic snapshot of an instance or a disk from one AWS Region to another in Amazon Lightsail.

When copying a manual snapshot, be sure to define the source region, source snapshot name, and target snapshot name parameters.

When copying an automatic snapshot, be sure to define the source region, source resource name, target snapshot name, and either the restore date or the use latest restorable auto snapshot parameters.

Creating a Request

data CopySnapshot

See: newCopySnapshot smart constructor.

Constructors

CopySnapshot'

Fields

useLatestRestorableAutoSnapshot :: Maybe Bool

A Boolean value to indicate whether to use the latest available automatic snapshot of the specified source instance or disk.

Constraints:

This parameter cannot be defined together with the restore date parameter. The use latest restorable auto snapshot and restore date parameters are mutually exclusive.

Define this parameter only when copying an automatic snapshot as a manual snapshot. For more information, see the Amazon Lightsail Developer Guide.

restoreDate :: Maybe Text

The date of the source automatic snapshot to copy. Use the get auto snapshots operation to identify the dates of the available automatic snapshots.

Constraints:

Must be specified in YYYY-MM-DD format.

This parameter cannot be defined together with the use latest restorable auto snapshot parameter. The restore date and use latest restorable auto snapshot parameters are mutually exclusive.

Define this parameter only when copying an automatic snapshot as a manual snapshot. For more information, see the Amazon Lightsail Developer Guide.

sourceResourceName :: Maybe Text

The name of the source instance or disk from which the source automatic snapshot was created.

Constraint:

Define this parameter only when copying an automatic snapshot as a manual snapshot. For more information, see the Amazon Lightsail Developer Guide.

sourceSnapshotName :: Maybe Text

The name of the source manual snapshot to copy.

Constraint:

Define this parameter only when copying a manual snapshot as another manual snapshot.

targetSnapshotName :: Text

The name of the new manual snapshot to be created as a copy.

sourceRegion :: RegionName

The AWS Region where the source manual or automatic snapshot is located.

newCopySnapshot

Arguments

:: Text
$sel:targetSnapshotName:CopySnapshot'

-> RegionName
$sel:sourceRegion:CopySnapshot'

-> CopySnapshot

Create a value of CopySnapshot with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:useLatestRestorableAutoSnapshot:CopySnapshot', copySnapshot_useLatestRestorableAutoSnapshot - A Boolean value to indicate whether to use the latest available automatic snapshot of the specified source instance or disk.

Constraints:

This parameter cannot be defined together with the restore date parameter. The use latest restorable auto snapshot and restore date parameters are mutually exclusive.

Define this parameter only when copying an automatic snapshot as a manual snapshot. For more information, see the Amazon Lightsail Developer Guide.

$sel:restoreDate:CopySnapshot', copySnapshot_restoreDate - The date of the source automatic snapshot to copy. Use the get auto snapshots operation to identify the dates of the available automatic snapshots.

Constraints:

Must be specified in YYYY-MM-DD format.

This parameter cannot be defined together with the use latest restorable auto snapshot parameter. The restore date and use latest restorable auto snapshot parameters are mutually exclusive.

Define this parameter only when copying an automatic snapshot as a manual snapshot. For more information, see the Amazon Lightsail Developer Guide.

$sel:sourceResourceName:CopySnapshot', copySnapshot_sourceResourceName - The name of the source instance or disk from which the source automatic snapshot was created.

Constraint:

Define this parameter only when copying an automatic snapshot as a manual snapshot. For more information, see the Amazon Lightsail Developer Guide.

$sel:sourceSnapshotName:CopySnapshot', copySnapshot_sourceSnapshotName - The name of the source manual snapshot to copy.

Constraint:

Define this parameter only when copying a manual snapshot as another manual snapshot.

$sel:targetSnapshotName:CopySnapshot', copySnapshot_targetSnapshotName - The name of the new manual snapshot to be created as a copy.

$sel:sourceRegion:CopySnapshot', copySnapshot_sourceRegion - The AWS Region where the source manual or automatic snapshot is located.

Request Lenses

copySnapshot_useLatestRestorableAutoSnapshot :: Lens' CopySnapshot (Maybe Bool)

A Boolean value to indicate whether to use the latest available automatic snapshot of the specified source instance or disk.

Constraints:

This parameter cannot be defined together with the restore date parameter. The use latest restorable auto snapshot and restore date parameters are mutually exclusive.

Define this parameter only when copying an automatic snapshot as a manual snapshot. For more information, see the Amazon Lightsail Developer Guide.

copySnapshot_restoreDate :: Lens' CopySnapshot (Maybe Text)

The date of the source automatic snapshot to copy. Use the get auto snapshots operation to identify the dates of the available automatic snapshots.

Constraints:

Must be specified in YYYY-MM-DD format.

This parameter cannot be defined together with the use latest restorable auto snapshot parameter. The restore date and use latest restorable auto snapshot parameters are mutually exclusive.

Define this parameter only when copying an automatic snapshot as a manual snapshot. For more information, see the Amazon Lightsail Developer Guide.

copySnapshot_sourceResourceName :: Lens' CopySnapshot (Maybe Text)

The name of the source instance or disk from which the source automatic snapshot was created.

Constraint:

Define this parameter only when copying an automatic snapshot as a manual snapshot. For more information, see the Amazon Lightsail Developer Guide.

copySnapshot_sourceSnapshotName :: Lens' CopySnapshot (Maybe Text)

The name of the source manual snapshot to copy.

Constraint:

Define this parameter only when copying a manual snapshot as another manual snapshot.

copySnapshot_targetSnapshotName :: Lens' CopySnapshot Text

The name of the new manual snapshot to be created as a copy.

copySnapshot_sourceRegion :: Lens' CopySnapshot RegionName

The AWS Region where the source manual or automatic snapshot is located.

Destructuring the Response

data CopySnapshotResponse

See: newCopySnapshotResponse smart constructor.

Constructors

CopySnapshotResponse'

Fields

operations :: Maybe [Operation]

An array of objects that describe the result of the action, such as the status of the request, the timestamp of the request, and the resources affected by the request.

httpStatus :: Int

The response's http status code.

newCopySnapshotResponse

Arguments

:: Int
$sel:httpStatus:CopySnapshotResponse'

-> CopySnapshotResponse

Create a value of CopySnapshotResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:operations:CopySnapshotResponse', copySnapshotResponse_operations - An array of objects that describe the result of the action, such as the status of the request, the timestamp of the request, and the resources affected by the request.

$sel:httpStatus:CopySnapshotResponse', copySnapshotResponse_httpStatus - The response's http status code.

Response Lenses

copySnapshotResponse_operations :: Lens' CopySnapshotResponse (Maybe [Operation])

An array of objects that describe the result of the action, such as the status of the request, the timestamp of the request, and the resources affected by the request.

copySnapshotResponse_httpStatus :: Lens' CopySnapshotResponse Int

The response's http status code.