Amazonka.Lightsail.CreateCloudFormationStack

Description

Creates an AWS CloudFormation stack, which creates a new Amazon EC2 instance from an exported Amazon Lightsail snapshot. This operation results in a CloudFormation stack record that can be used to track the AWS CloudFormation stack created. Use the get cloud formation stack records operation to get a list of the CloudFormation stacks created.

Wait until after your new Amazon EC2 instance is created before running the create cloud formation stack operation again with the same export snapshot record.

Creating a Request

data CreateCloudFormationStack

See: newCreateCloudFormationStack smart constructor.

Constructors

CreateCloudFormationStack'

Fields

instances :: [InstanceEntry]

An array of parameters that will be used to create the new Amazon EC2 instance. You can only pass one instance entry at a time in this array. You will get an invalid parameter error if you pass more than one instance entry in this array.

newCreateCloudFormationStack :: CreateCloudFormationStack

Create a value of CreateCloudFormationStack with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:instances:CreateCloudFormationStack', createCloudFormationStack_instances - An array of parameters that will be used to create the new Amazon EC2 instance. You can only pass one instance entry at a time in this array. You will get an invalid parameter error if you pass more than one instance entry in this array.

Request Lenses

createCloudFormationStack_instances :: Lens' CreateCloudFormationStack [InstanceEntry]

An array of parameters that will be used to create the new Amazon EC2 instance. You can only pass one instance entry at a time in this array. You will get an invalid parameter error if you pass more than one instance entry in this array.

Destructuring the Response

data CreateCloudFormationStackResponse

See: newCreateCloudFormationStackResponse smart constructor.

Constructors

CreateCloudFormationStackResponse'

Fields

operations :: Maybe [Operation]

An array of objects that describe the result of the action, such as the status of the request, the timestamp of the request, and the resources affected by the request.

httpStatus :: Int

The response's http status code.

newCreateCloudFormationStackResponse

Arguments

:: Int
$sel:httpStatus:CreateCloudFormationStackResponse'

-> CreateCloudFormationStackResponse

Create a value of CreateCloudFormationStackResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:operations:CreateCloudFormationStackResponse', createCloudFormationStackResponse_operations - An array of objects that describe the result of the action, such as the status of the request, the timestamp of the request, and the resources affected by the request.

$sel:httpStatus:CreateCloudFormationStackResponse', createCloudFormationStackResponse_httpStatus - The response's http status code.

Response Lenses

createCloudFormationStackResponse_operations :: Lens' CreateCloudFormationStackResponse (Maybe [Operation])

An array of objects that describe the result of the action, such as the status of the request, the timestamp of the request, and the resources affected by the request.

createCloudFormationStackResponse_httpStatus :: Lens' CreateCloudFormationStackResponse Int

The response's http status code.