Amazonka.Lightsail.CreateDisk

Description

Creates a block storage disk that can be attached to an Amazon Lightsail instance in the same Availability Zone (e.g., us-east-2a).

The create disk operation supports tag-based access control via request tags. For more information, see the Amazon Lightsail Developer Guide.

Creating a Request

data CreateDisk

See: newCreateDisk smart constructor.

Constructors

CreateDisk'

Fields

addOns :: Maybe [AddOnRequest]

An array of objects that represent the add-ons to enable for the new disk.

tags :: Maybe [Tag]

The tag keys and optional values to add to the resource during create.

Use the TagResource action to tag a resource after it's created.

diskName :: Text

The unique Lightsail disk name (e.g., my-disk).

availabilityZone :: Text

The Availability Zone where you want to create the disk (e.g., us-east-2a). Use the same Availability Zone as the Lightsail instance to which you want to attach the disk.

Use the get regions operation to list the Availability Zones where Lightsail is currently available.

sizeInGb :: Int

The size of the disk in GB (e.g., 32).

newCreateDisk

Arguments

:: Text
$sel:diskName:CreateDisk'

-> Text
$sel:availabilityZone:CreateDisk'

-> Int
$sel:sizeInGb:CreateDisk'

-> CreateDisk

Create a value of CreateDisk with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:addOns:CreateDisk', createDisk_addOns - An array of objects that represent the add-ons to enable for the new disk.

$sel:tags:CreateDisk', createDisk_tags - The tag keys and optional values to add to the resource during create.

Use the TagResource action to tag a resource after it's created.

$sel:diskName:CreateDisk', createDisk_diskName - The unique Lightsail disk name (e.g., my-disk).

$sel:availabilityZone:CreateDisk', createDisk_availabilityZone - The Availability Zone where you want to create the disk (e.g., us-east-2a). Use the same Availability Zone as the Lightsail instance to which you want to attach the disk.

Use the get regions operation to list the Availability Zones where Lightsail is currently available.

$sel:sizeInGb:CreateDisk', createDisk_sizeInGb - The size of the disk in GB (e.g., 32).

Request Lenses

createDisk_addOns :: Lens' CreateDisk (Maybe [AddOnRequest])

An array of objects that represent the add-ons to enable for the new disk.

createDisk_tags :: Lens' CreateDisk (Maybe [Tag])

The tag keys and optional values to add to the resource during create.

Use the TagResource action to tag a resource after it's created.

createDisk_diskName :: Lens' CreateDisk Text

The unique Lightsail disk name (e.g., my-disk).

createDisk_availabilityZone :: Lens' CreateDisk Text

The Availability Zone where you want to create the disk (e.g., us-east-2a). Use the same Availability Zone as the Lightsail instance to which you want to attach the disk.

Use the get regions operation to list the Availability Zones where Lightsail is currently available.

createDisk_sizeInGb :: Lens' CreateDisk Int

The size of the disk in GB (e.g., 32).

Destructuring the Response

data CreateDiskResponse

See: newCreateDiskResponse smart constructor.

Constructors

CreateDiskResponse'

Fields

operations :: Maybe [Operation]

An array of objects that describe the result of the action, such as the status of the request, the timestamp of the request, and the resources affected by the request.

httpStatus :: Int

The response's http status code.

newCreateDiskResponse

Arguments

:: Int
$sel:httpStatus:CreateDiskResponse'

-> CreateDiskResponse

Create a value of CreateDiskResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:operations:CreateDiskResponse', createDiskResponse_operations - An array of objects that describe the result of the action, such as the status of the request, the timestamp of the request, and the resources affected by the request.

$sel:httpStatus:CreateDiskResponse', createDiskResponse_httpStatus - The response's http status code.

Response Lenses

createDiskResponse_operations :: Lens' CreateDiskResponse (Maybe [Operation])

An array of objects that describe the result of the action, such as the status of the request, the timestamp of the request, and the resources affected by the request.

createDiskResponse_httpStatus :: Lens' CreateDiskResponse Int

The response's http status code.