Amazonka.Lightsail.CreateLoadBalancerTlsCertificate

Description

Creates an SSL/TLS certificate for an Amazon Lightsail load balancer.

TLS is just an updated, more secure version of Secure Socket Layer (SSL).

The CreateLoadBalancerTlsCertificate operation supports tag-based access control via resource tags applied to the resource identified by load balancer name. For more information, see the Amazon Lightsail Developer Guide.

Creating a Request

data CreateLoadBalancerTlsCertificate

See: newCreateLoadBalancerTlsCertificate smart constructor.

Constructors

CreateLoadBalancerTlsCertificate'

Fields

certificateAlternativeNames :: Maybe [Text]

An array of strings listing alternative domains and subdomains for your SSL/TLS certificate. Lightsail will de-dupe the names for you. You can have a maximum of 9 alternative names (in addition to the 1 primary domain). We do not support wildcards (e.g., *.example.com).

tags :: Maybe [Tag]

The tag keys and optional values to add to the resource during create.

Use the TagResource action to tag a resource after it's created.

loadBalancerName :: Text

The load balancer name where you want to create the SSL/TLS certificate.

certificateName :: Text

The SSL/TLS certificate name.

You can have up to 10 certificates in your account at one time. Each Lightsail load balancer can have up to 2 certificates associated with it at one time. There is also an overall limit to the number of certificates that can be issue in a 365-day period. For more information, see Limits.

certificateDomainName :: Text

The domain name (e.g., example.com) for your SSL/TLS certificate.

newCreateLoadBalancerTlsCertificate

Arguments

:: Text
$sel:loadBalancerName:CreateLoadBalancerTlsCertificate'

-> Text
$sel:certificateName:CreateLoadBalancerTlsCertificate'

-> Text
$sel:certificateDomainName:CreateLoadBalancerTlsCertificate'

-> CreateLoadBalancerTlsCertificate

Create a value of CreateLoadBalancerTlsCertificate with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:certificateAlternativeNames:CreateLoadBalancerTlsCertificate', createLoadBalancerTlsCertificate_certificateAlternativeNames - An array of strings listing alternative domains and subdomains for your SSL/TLS certificate. Lightsail will de-dupe the names for you. You can have a maximum of 9 alternative names (in addition to the 1 primary domain). We do not support wildcards (e.g., *.example.com).

$sel:tags:CreateLoadBalancerTlsCertificate', createLoadBalancerTlsCertificate_tags - The tag keys and optional values to add to the resource during create.

Use the TagResource action to tag a resource after it's created.

$sel:loadBalancerName:CreateLoadBalancerTlsCertificate', createLoadBalancerTlsCertificate_loadBalancerName - The load balancer name where you want to create the SSL/TLS certificate.

$sel:certificateName:CreateLoadBalancerTlsCertificate', createLoadBalancerTlsCertificate_certificateName - The SSL/TLS certificate name.

You can have up to 10 certificates in your account at one time. Each Lightsail load balancer can have up to 2 certificates associated with it at one time. There is also an overall limit to the number of certificates that can be issue in a 365-day period. For more information, see Limits.

$sel:certificateDomainName:CreateLoadBalancerTlsCertificate', createLoadBalancerTlsCertificate_certificateDomainName - The domain name (e.g., example.com) for your SSL/TLS certificate.

Request Lenses

createLoadBalancerTlsCertificate_certificateAlternativeNames :: Lens' CreateLoadBalancerTlsCertificate (Maybe [Text])

An array of strings listing alternative domains and subdomains for your SSL/TLS certificate. Lightsail will de-dupe the names for you. You can have a maximum of 9 alternative names (in addition to the 1 primary domain). We do not support wildcards (e.g., *.example.com).

createLoadBalancerTlsCertificate_tags :: Lens' CreateLoadBalancerTlsCertificate (Maybe [Tag])

The tag keys and optional values to add to the resource during create.

Use the TagResource action to tag a resource after it's created.

createLoadBalancerTlsCertificate_loadBalancerName :: Lens' CreateLoadBalancerTlsCertificate Text

The load balancer name where you want to create the SSL/TLS certificate.

createLoadBalancerTlsCertificate_certificateName :: Lens' CreateLoadBalancerTlsCertificate Text

The SSL/TLS certificate name.

You can have up to 10 certificates in your account at one time. Each Lightsail load balancer can have up to 2 certificates associated with it at one time. There is also an overall limit to the number of certificates that can be issue in a 365-day period. For more information, see Limits.

createLoadBalancerTlsCertificate_certificateDomainName :: Lens' CreateLoadBalancerTlsCertificate Text

The domain name (e.g., example.com) for your SSL/TLS certificate.

Destructuring the Response

data CreateLoadBalancerTlsCertificateResponse

See: newCreateLoadBalancerTlsCertificateResponse smart constructor.

Constructors

CreateLoadBalancerTlsCertificateResponse'

Fields

operations :: Maybe [Operation]

An array of objects that describe the result of the action, such as the status of the request, the timestamp of the request, and the resources affected by the request.

httpStatus :: Int

The response's http status code.

newCreateLoadBalancerTlsCertificateResponse

Arguments

:: Int
$sel:httpStatus:CreateLoadBalancerTlsCertificateResponse'

-> CreateLoadBalancerTlsCertificateResponse

Create a value of CreateLoadBalancerTlsCertificateResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:operations:CreateLoadBalancerTlsCertificateResponse', createLoadBalancerTlsCertificateResponse_operations - An array of objects that describe the result of the action, such as the status of the request, the timestamp of the request, and the resources affected by the request.

$sel:httpStatus:CreateLoadBalancerTlsCertificateResponse', createLoadBalancerTlsCertificateResponse_httpStatus - The response's http status code.

Response Lenses

createLoadBalancerTlsCertificateResponse_operations :: Lens' CreateLoadBalancerTlsCertificateResponse (Maybe [Operation])

An array of objects that describe the result of the action, such as the status of the request, the timestamp of the request, and the resources affected by the request.

createLoadBalancerTlsCertificateResponse_httpStatus :: Lens' CreateLoadBalancerTlsCertificateResponse Int

The response's http status code.