Amazonka.Lightsail.DeleteLoadBalancerTlsCertificate

Description

Deletes an SSL/TLS certificate associated with a Lightsail load balancer.

The DeleteLoadBalancerTlsCertificate operation supports tag-based access control via resource tags applied to the resource identified by load balancer name. For more information, see the Amazon Lightsail Developer Guide.

Creating a Request

data DeleteLoadBalancerTlsCertificate

See: newDeleteLoadBalancerTlsCertificate smart constructor.

Constructors

DeleteLoadBalancerTlsCertificate'

Fields

force :: Maybe Bool

When true, forces the deletion of an SSL/TLS certificate.

There can be two certificates associated with a Lightsail load balancer: the primary and the backup. The force parameter is required when the primary SSL/TLS certificate is in use by an instance attached to the load balancer.

loadBalancerName :: Text

The load balancer name.

certificateName :: Text

The SSL/TLS certificate name.

newDeleteLoadBalancerTlsCertificate

Arguments

:: Text
$sel:loadBalancerName:DeleteLoadBalancerTlsCertificate'

-> Text
$sel:certificateName:DeleteLoadBalancerTlsCertificate'

-> DeleteLoadBalancerTlsCertificate

Create a value of DeleteLoadBalancerTlsCertificate with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:force:DeleteLoadBalancerTlsCertificate', deleteLoadBalancerTlsCertificate_force - When true, forces the deletion of an SSL/TLS certificate.

There can be two certificates associated with a Lightsail load balancer: the primary and the backup. The force parameter is required when the primary SSL/TLS certificate is in use by an instance attached to the load balancer.

$sel:loadBalancerName:DeleteLoadBalancerTlsCertificate', deleteLoadBalancerTlsCertificate_loadBalancerName - The load balancer name.

$sel:certificateName:DeleteLoadBalancerTlsCertificate', deleteLoadBalancerTlsCertificate_certificateName - The SSL/TLS certificate name.

Request Lenses

deleteLoadBalancerTlsCertificate_force :: Lens' DeleteLoadBalancerTlsCertificate (Maybe Bool)

When true, forces the deletion of an SSL/TLS certificate.

There can be two certificates associated with a Lightsail load balancer: the primary and the backup. The force parameter is required when the primary SSL/TLS certificate is in use by an instance attached to the load balancer.

deleteLoadBalancerTlsCertificate_loadBalancerName :: Lens' DeleteLoadBalancerTlsCertificate Text

The load balancer name.

deleteLoadBalancerTlsCertificate_certificateName :: Lens' DeleteLoadBalancerTlsCertificate Text

The SSL/TLS certificate name.

Destructuring the Response

data DeleteLoadBalancerTlsCertificateResponse

See: newDeleteLoadBalancerTlsCertificateResponse smart constructor.

Constructors

DeleteLoadBalancerTlsCertificateResponse'

Fields

operations :: Maybe [Operation]

An array of objects that describe the result of the action, such as the status of the request, the timestamp of the request, and the resources affected by the request.

httpStatus :: Int

The response's http status code.

newDeleteLoadBalancerTlsCertificateResponse

Arguments

:: Int
$sel:httpStatus:DeleteLoadBalancerTlsCertificateResponse'

-> DeleteLoadBalancerTlsCertificateResponse

Create a value of DeleteLoadBalancerTlsCertificateResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:operations:DeleteLoadBalancerTlsCertificateResponse', deleteLoadBalancerTlsCertificateResponse_operations - An array of objects that describe the result of the action, such as the status of the request, the timestamp of the request, and the resources affected by the request.

$sel:httpStatus:DeleteLoadBalancerTlsCertificateResponse', deleteLoadBalancerTlsCertificateResponse_httpStatus - The response's http status code.

Response Lenses

deleteLoadBalancerTlsCertificateResponse_operations :: Lens' DeleteLoadBalancerTlsCertificateResponse (Maybe [Operation])

An array of objects that describe the result of the action, such as the status of the request, the timestamp of the request, and the resources affected by the request.

deleteLoadBalancerTlsCertificateResponse_httpStatus :: Lens' DeleteLoadBalancerTlsCertificateResponse Int

The response's http status code.