Amazonka.Lightsail.GetBucketMetricData

Description

Returns the data points of a specific metric for an Amazon Lightsail bucket.

Metrics report the utilization of a bucket. View and collect metric data regularly to monitor the number of objects stored in a bucket (including object versions) and the storage space used by those objects.

Creating a Request

data GetBucketMetricData

See: newGetBucketMetricData smart constructor.

Constructors

GetBucketMetricData'

Fields

bucketName :: Text

The name of the bucket for which to get metric data.

metricName :: BucketMetricName

The metric for which you want to return information.

Valid bucket metric names are listed below, along with the most useful statistics to include in your request, and the published unit value.

These bucket metrics are reported once per day.

BucketSizeBytes - The amount of data in bytes stored in a bucket. This value is calculated by summing the size of all objects in the bucket (including object versions), including the size of all parts for all incomplete multipart uploads to the bucket.

Statistics: The most useful statistic is Maximum.

Unit: The published unit is Bytes.

NumberOfObjects - The total number of objects stored in a bucket. This value is calculated by counting all objects in the bucket (including object versions) and the total number of parts for all incomplete multipart uploads to the bucket.

Statistics: The most useful statistic is Average.

Unit: The published unit is Count.

startTime :: POSIX

The timestamp indicating the earliest data to be returned.

endTime :: POSIX

The timestamp indicating the latest data to be returned.

period :: Natural

The granularity, in seconds, of the returned data points.

Bucket storage metrics are reported once per day. Therefore, you should specify a period of 86400 seconds, which is the number of seconds in a day.

statistics :: [MetricStatistic]

The statistic for the metric.

The following statistics are available:

Minimum - The lowest value observed during the specified period. Use this value to determine low volumes of activity for your application.

Maximum - The highest value observed during the specified period. Use this value to determine high volumes of activity for your application.

Sum - The sum of all values submitted for the matching metric. You can use this statistic to determine the total volume of a metric.

Average - The value of Sum / SampleCount during the specified period. By comparing this statistic with the Minimum and Maximum values, you can determine the full scope of a metric and how close the average use is to the Minimum and Maximum values. This comparison helps you to know when to increase or decrease your resources.

SampleCount - The count, or number, of data points used for the statistical calculation.

unit :: MetricUnit

The unit for the metric data request.

Valid units depend on the metric data being requested. For the valid units with each available metric, see the metricName parameter.

newGetBucketMetricData

Arguments

:: Text
$sel:bucketName:GetBucketMetricData'

-> BucketMetricName
$sel:metricName:GetBucketMetricData'

-> UTCTime
$sel:startTime:GetBucketMetricData'

-> UTCTime
$sel:endTime:GetBucketMetricData'

-> Natural
$sel:period:GetBucketMetricData'

-> MetricUnit
$sel:unit:GetBucketMetricData'

-> GetBucketMetricData

Create a value of GetBucketMetricData with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:bucketName:GetBucketMetricData', getBucketMetricData_bucketName - The name of the bucket for which to get metric data.

$sel:metricName:GetBucketMetricData', getBucketMetricData_metricName - The metric for which you want to return information.

Valid bucket metric names are listed below, along with the most useful statistics to include in your request, and the published unit value.

These bucket metrics are reported once per day.

BucketSizeBytes - The amount of data in bytes stored in a bucket. This value is calculated by summing the size of all objects in the bucket (including object versions), including the size of all parts for all incomplete multipart uploads to the bucket.

Statistics: The most useful statistic is Maximum.

Unit: The published unit is Bytes.

NumberOfObjects - The total number of objects stored in a bucket. This value is calculated by counting all objects in the bucket (including object versions) and the total number of parts for all incomplete multipart uploads to the bucket.

Statistics: The most useful statistic is Average.

Unit: The published unit is Count.

$sel:startTime:GetBucketMetricData', getBucketMetricData_startTime - The timestamp indicating the earliest data to be returned.

$sel:endTime:GetBucketMetricData', getBucketMetricData_endTime - The timestamp indicating the latest data to be returned.

$sel:period:GetBucketMetricData', getBucketMetricData_period - The granularity, in seconds, of the returned data points.

Bucket storage metrics are reported once per day. Therefore, you should specify a period of 86400 seconds, which is the number of seconds in a day.

$sel:statistics:GetBucketMetricData', getBucketMetricData_statistics - The statistic for the metric.

The following statistics are available:

Minimum - The lowest value observed during the specified period. Use this value to determine low volumes of activity for your application.

Maximum - The highest value observed during the specified period. Use this value to determine high volumes of activity for your application.

Sum - The sum of all values submitted for the matching metric. You can use this statistic to determine the total volume of a metric.

Average - The value of Sum / SampleCount during the specified period. By comparing this statistic with the Minimum and Maximum values, you can determine the full scope of a metric and how close the average use is to the Minimum and Maximum values. This comparison helps you to know when to increase or decrease your resources.

SampleCount - The count, or number, of data points used for the statistical calculation.

$sel:unit:GetBucketMetricData', getBucketMetricData_unit - The unit for the metric data request.

Valid units depend on the metric data being requested. For the valid units with each available metric, see the metricName parameter.

Request Lenses

getBucketMetricData_bucketName :: Lens' GetBucketMetricData Text

The name of the bucket for which to get metric data.

getBucketMetricData_metricName :: Lens' GetBucketMetricData BucketMetricName

The metric for which you want to return information.

Valid bucket metric names are listed below, along with the most useful statistics to include in your request, and the published unit value.

These bucket metrics are reported once per day.

BucketSizeBytes - The amount of data in bytes stored in a bucket. This value is calculated by summing the size of all objects in the bucket (including object versions), including the size of all parts for all incomplete multipart uploads to the bucket.

Statistics: The most useful statistic is Maximum.

Unit: The published unit is Bytes.

NumberOfObjects - The total number of objects stored in a bucket. This value is calculated by counting all objects in the bucket (including object versions) and the total number of parts for all incomplete multipart uploads to the bucket.

Statistics: The most useful statistic is Average.

Unit: The published unit is Count.

getBucketMetricData_startTime :: Lens' GetBucketMetricData UTCTime

The timestamp indicating the earliest data to be returned.

getBucketMetricData_endTime :: Lens' GetBucketMetricData UTCTime

The timestamp indicating the latest data to be returned.

getBucketMetricData_period :: Lens' GetBucketMetricData Natural

The granularity, in seconds, of the returned data points.

Bucket storage metrics are reported once per day. Therefore, you should specify a period of 86400 seconds, which is the number of seconds in a day.

getBucketMetricData_statistics :: Lens' GetBucketMetricData [MetricStatistic]

The statistic for the metric.

The following statistics are available:

Minimum - The lowest value observed during the specified period. Use this value to determine low volumes of activity for your application.

Maximum - The highest value observed during the specified period. Use this value to determine high volumes of activity for your application.

Sum - The sum of all values submitted for the matching metric. You can use this statistic to determine the total volume of a metric.

Average - The value of Sum / SampleCount during the specified period. By comparing this statistic with the Minimum and Maximum values, you can determine the full scope of a metric and how close the average use is to the Minimum and Maximum values. This comparison helps you to know when to increase or decrease your resources.

SampleCount - The count, or number, of data points used for the statistical calculation.

getBucketMetricData_unit :: Lens' GetBucketMetricData MetricUnit

The unit for the metric data request.

Valid units depend on the metric data being requested. For the valid units with each available metric, see the metricName parameter.

Destructuring the Response

data GetBucketMetricDataResponse

See: newGetBucketMetricDataResponse smart constructor.

Constructors

GetBucketMetricDataResponse'

Fields

metricName :: Maybe BucketMetricName

The name of the metric returned.

metricData :: Maybe [MetricDatapoint]

An array of objects that describe the metric data returned.

httpStatus :: Int

The response's http status code.

newGetBucketMetricDataResponse

Arguments

:: Int
$sel:httpStatus:GetBucketMetricDataResponse'

-> GetBucketMetricDataResponse

Create a value of GetBucketMetricDataResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:metricName:GetBucketMetricData', getBucketMetricDataResponse_metricName - The name of the metric returned.

$sel:metricData:GetBucketMetricDataResponse', getBucketMetricDataResponse_metricData - An array of objects that describe the metric data returned.

$sel:httpStatus:GetBucketMetricDataResponse', getBucketMetricDataResponse_httpStatus - The response's http status code.

Response Lenses

getBucketMetricDataResponse_metricName :: Lens' GetBucketMetricDataResponse (Maybe BucketMetricName)

The name of the metric returned.

getBucketMetricDataResponse_metricData :: Lens' GetBucketMetricDataResponse (Maybe [MetricDatapoint])

An array of objects that describe the metric data returned.

getBucketMetricDataResponse_httpStatus :: Lens' GetBucketMetricDataResponse Int

The response's http status code.