Amazonka.Lightsail.GetCertificates

Description

Returns information about one or more Amazon Lightsail SSL/TLS certificates.

To get a summary of a certificate, ommit includeCertificateDetails from your request. The response will include only the certificate Amazon Resource Name (ARN), certificate name, domain name, and tags.

Creating a Request

data GetCertificates

See: newGetCertificates smart constructor.

Constructors

GetCertificates'

Fields

certificateStatuses :: Maybe [CertificateStatus]

The status of the certificates for which to return information.

For example, specify ISSUED to return only certificates with an ISSUED status.

When omitted, the response includes all of your certificates in the AWS Region where the request is made, regardless of their current status.

certificateName :: Maybe Text

The name for the certificate for which to return information.

When omitted, the response includes all of your certificates in the AWS Region where the request is made.

includeCertificateDetails :: Maybe Bool

Indicates whether to include detailed information about the certificates in the response.

When omitted, the response includes only the certificate names, Amazon Resource Names (ARNs), domain names, and tags.

newGetCertificates :: GetCertificates

Create a value of GetCertificates with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:certificateStatuses:GetCertificates', getCertificates_certificateStatuses - The status of the certificates for which to return information.

For example, specify ISSUED to return only certificates with an ISSUED status.

When omitted, the response includes all of your certificates in the AWS Region where the request is made, regardless of their current status.

$sel:certificateName:GetCertificates', getCertificates_certificateName - The name for the certificate for which to return information.

When omitted, the response includes all of your certificates in the AWS Region where the request is made.

$sel:includeCertificateDetails:GetCertificates', getCertificates_includeCertificateDetails - Indicates whether to include detailed information about the certificates in the response.

When omitted, the response includes only the certificate names, Amazon Resource Names (ARNs), domain names, and tags.

Request Lenses

getCertificates_certificateStatuses :: Lens' GetCertificates (Maybe [CertificateStatus])

The status of the certificates for which to return information.

For example, specify ISSUED to return only certificates with an ISSUED status.

When omitted, the response includes all of your certificates in the AWS Region where the request is made, regardless of their current status.

getCertificates_certificateName :: Lens' GetCertificates (Maybe Text)

The name for the certificate for which to return information.

When omitted, the response includes all of your certificates in the AWS Region where the request is made.

getCertificates_includeCertificateDetails :: Lens' GetCertificates (Maybe Bool)

Indicates whether to include detailed information about the certificates in the response.

When omitted, the response includes only the certificate names, Amazon Resource Names (ARNs), domain names, and tags.

Destructuring the Response

data GetCertificatesResponse

See: newGetCertificatesResponse smart constructor.

Constructors

GetCertificatesResponse'

Fields

certificates :: Maybe [CertificateSummary]

An object that describes certificates.

httpStatus :: Int

The response's http status code.

newGetCertificatesResponse

Arguments

:: Int
$sel:httpStatus:GetCertificatesResponse'

-> GetCertificatesResponse

Create a value of GetCertificatesResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:certificates:GetCertificatesResponse', getCertificatesResponse_certificates - An object that describes certificates.

$sel:httpStatus:GetCertificatesResponse', getCertificatesResponse_httpStatus - The response's http status code.

Response Lenses

getCertificatesResponse_certificates :: Lens' GetCertificatesResponse (Maybe [CertificateSummary])

An object that describes certificates.

getCertificatesResponse_httpStatus :: Lens' GetCertificatesResponse Int

The response's http status code.