Amazonka.Lightsail.GetContainerLog

Description

Returns the log events of a container of your Amazon Lightsail container service.

If your container service has more than one node (i.e., a scale greater than 1), then the log events that are returned for the specified container are merged from all nodes on your container service.

Container logs are retained for a certain amount of time. For more information, see Amazon Lightsail endpoints and quotas in the AWS General Reference.

Creating a Request

data GetContainerLog

See: newGetContainerLog smart constructor.

Constructors

GetContainerLog'

Fields

startTime :: Maybe POSIX

The start of the time interval for which to get log data.

Constraints:

Specified in Coordinated Universal Time (UTC).

Specified in the Unix time format.

For example, if you wish to use a start time of October 1, 2018, at 8 PM UTC, specify 1538424000 as the start time.

You can convert a human-friendly time to Unix time format using a converter like Epoch converter.

endTime :: Maybe POSIX

The end of the time interval for which to get log data.

Constraints:

Specified in Coordinated Universal Time (UTC).

Specified in the Unix time format.

For example, if you wish to use an end time of October 1, 2018, at 9 PM UTC, specify 1538427600 as the end time.

You can convert a human-friendly time to Unix time format using a converter like Epoch converter.

pageToken :: Maybe Text

The token to advance to the next page of results from your request.

To get a page token, perform an initial GetContainerLog request. If your results are paginated, the response will return a next page token that you can specify as the page token in a subsequent request.

filterPattern :: Maybe Text

The pattern to use to filter the returned log events to a specific term.

The following are a few examples of filter patterns that you can specify:

To return all log events, specify a filter pattern of "".

To exclude log events that contain the ERROR term, and return all other log events, specify a filter pattern of "-ERROR".

To return log events that contain the ERROR term, specify a filter pattern of "ERROR".

To return log events that contain both the ERROR and Exception terms, specify a filter pattern of "ERROR Exception".

To return log events that contain the ERROR or the Exception term, specify a filter pattern of "?ERROR ?Exception".

serviceName :: Text

The name of the container service for which to get a container log.

containerName :: Text

The name of the container that is either running or previously ran on the container service for which to return a log.

newGetContainerLog

Arguments

:: Text
$sel:serviceName:GetContainerLog'

-> Text
$sel:containerName:GetContainerLog'

-> GetContainerLog

Create a value of GetContainerLog with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:startTime:GetContainerLog', getContainerLog_startTime - The start of the time interval for which to get log data.

Constraints:

Specified in Coordinated Universal Time (UTC).

Specified in the Unix time format.

For example, if you wish to use a start time of October 1, 2018, at 8 PM UTC, specify 1538424000 as the start time.

You can convert a human-friendly time to Unix time format using a converter like Epoch converter.

$sel:endTime:GetContainerLog', getContainerLog_endTime - The end of the time interval for which to get log data.

Constraints:

Specified in Coordinated Universal Time (UTC).

Specified in the Unix time format.

For example, if you wish to use an end time of October 1, 2018, at 9 PM UTC, specify 1538427600 as the end time.

You can convert a human-friendly time to Unix time format using a converter like Epoch converter.

$sel:pageToken:GetContainerLog', getContainerLog_pageToken - The token to advance to the next page of results from your request.

To get a page token, perform an initial GetContainerLog request. If your results are paginated, the response will return a next page token that you can specify as the page token in a subsequent request.

$sel:filterPattern:GetContainerLog', getContainerLog_filterPattern - The pattern to use to filter the returned log events to a specific term.

The following are a few examples of filter patterns that you can specify:

To return all log events, specify a filter pattern of "".

To exclude log events that contain the ERROR term, and return all other log events, specify a filter pattern of "-ERROR".

To return log events that contain the ERROR term, specify a filter pattern of "ERROR".

To return log events that contain both the ERROR and Exception terms, specify a filter pattern of "ERROR Exception".

To return log events that contain the ERROR or the Exception term, specify a filter pattern of "?ERROR ?Exception".

$sel:serviceName:GetContainerLog', getContainerLog_serviceName - The name of the container service for which to get a container log.

$sel:containerName:GetContainerLog', getContainerLog_containerName - The name of the container that is either running or previously ran on the container service for which to return a log.

Request Lenses

getContainerLog_startTime :: Lens' GetContainerLog (Maybe UTCTime)

The start of the time interval for which to get log data.

Constraints:

Specified in Coordinated Universal Time (UTC).

Specified in the Unix time format.

For example, if you wish to use a start time of October 1, 2018, at 8 PM UTC, specify 1538424000 as the start time.

You can convert a human-friendly time to Unix time format using a converter like Epoch converter.

getContainerLog_endTime :: Lens' GetContainerLog (Maybe UTCTime)

The end of the time interval for which to get log data.

Constraints:

Specified in Coordinated Universal Time (UTC).

Specified in the Unix time format.

For example, if you wish to use an end time of October 1, 2018, at 9 PM UTC, specify 1538427600 as the end time.

You can convert a human-friendly time to Unix time format using a converter like Epoch converter.

getContainerLog_pageToken :: Lens' GetContainerLog (Maybe Text)

The token to advance to the next page of results from your request.

To get a page token, perform an initial GetContainerLog request. If your results are paginated, the response will return a next page token that you can specify as the page token in a subsequent request.

getContainerLog_filterPattern :: Lens' GetContainerLog (Maybe Text)

The pattern to use to filter the returned log events to a specific term.

The following are a few examples of filter patterns that you can specify:

To return all log events, specify a filter pattern of "".

To exclude log events that contain the ERROR term, and return all other log events, specify a filter pattern of "-ERROR".

To return log events that contain the ERROR term, specify a filter pattern of "ERROR".

To return log events that contain both the ERROR and Exception terms, specify a filter pattern of "ERROR Exception".

To return log events that contain the ERROR or the Exception term, specify a filter pattern of "?ERROR ?Exception".

getContainerLog_serviceName :: Lens' GetContainerLog Text

The name of the container service for which to get a container log.

getContainerLog_containerName :: Lens' GetContainerLog Text

The name of the container that is either running or previously ran on the container service for which to return a log.

Destructuring the Response

data GetContainerLogResponse

See: newGetContainerLogResponse smart constructor.

Constructors

GetContainerLogResponse'

Fields

nextPageToken :: Maybe Text

The token to advance to the next page of results from your request.

A next page token is not returned if there are no more results to display.

To get the next page of results, perform another GetContainerLog request and specify the next page token using the pageToken parameter.

logEvents :: Maybe [ContainerServiceLogEvent]

An array of objects that describe the log events of a container.

httpStatus :: Int

The response's http status code.

newGetContainerLogResponse

Arguments

:: Int
$sel:httpStatus:GetContainerLogResponse'

-> GetContainerLogResponse

Create a value of GetContainerLogResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:nextPageToken:GetContainerLogResponse', getContainerLogResponse_nextPageToken - The token to advance to the next page of results from your request.

A next page token is not returned if there are no more results to display.

To get the next page of results, perform another GetContainerLog request and specify the next page token using the pageToken parameter.

$sel:logEvents:GetContainerLogResponse', getContainerLogResponse_logEvents - An array of objects that describe the log events of a container.

$sel:httpStatus:GetContainerLogResponse', getContainerLogResponse_httpStatus - The response's http status code.

Response Lenses

getContainerLogResponse_nextPageToken :: Lens' GetContainerLogResponse (Maybe Text)

The token to advance to the next page of results from your request.

A next page token is not returned if there are no more results to display.

To get the next page of results, perform another GetContainerLog request and specify the next page token using the pageToken parameter.

getContainerLogResponse_logEvents :: Lens' GetContainerLogResponse (Maybe [ContainerServiceLogEvent])

An array of objects that describe the log events of a container.

getContainerLogResponse_httpStatus :: Lens' GetContainerLogResponse Int

The response's http status code.