Amazonka.Lightsail.GetRelationalDatabase

Description

Returns information about a specific database in Amazon Lightsail.

Creating a Request

data GetRelationalDatabase

See: newGetRelationalDatabase smart constructor.

Constructors

GetRelationalDatabase'

Fields

relationalDatabaseName :: Text

The name of the database that you are looking up.

newGetRelationalDatabase

Arguments

:: Text
$sel:relationalDatabaseName:GetRelationalDatabase'

-> GetRelationalDatabase

Create a value of GetRelationalDatabase with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:relationalDatabaseName:GetRelationalDatabase', getRelationalDatabase_relationalDatabaseName - The name of the database that you are looking up.

Request Lenses

getRelationalDatabase_relationalDatabaseName :: Lens' GetRelationalDatabase Text

The name of the database that you are looking up.

Destructuring the Response

data GetRelationalDatabaseResponse

See: newGetRelationalDatabaseResponse smart constructor.

Constructors

GetRelationalDatabaseResponse'

Fields

relationalDatabase :: Maybe RelationalDatabase

An object describing the specified database.

httpStatus :: Int

The response's http status code.

newGetRelationalDatabaseResponse

Arguments

:: Int
$sel:httpStatus:GetRelationalDatabaseResponse'

-> GetRelationalDatabaseResponse

Create a value of GetRelationalDatabaseResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:relationalDatabase:GetRelationalDatabaseResponse', getRelationalDatabaseResponse_relationalDatabase - An object describing the specified database.

$sel:httpStatus:GetRelationalDatabaseResponse', getRelationalDatabaseResponse_httpStatus - The response's http status code.

Response Lenses

getRelationalDatabaseResponse_relationalDatabase :: Lens' GetRelationalDatabaseResponse (Maybe RelationalDatabase)

An object describing the specified database.

getRelationalDatabaseResponse_httpStatus :: Lens' GetRelationalDatabaseResponse Int

The response's http status code.