Amazonka.Lightsail.GetRelationalDatabaseLogEvents

Description

Returns a list of log events for a database in Amazon Lightsail.

Creating a Request

data GetRelationalDatabaseLogEvents

See: newGetRelationalDatabaseLogEvents smart constructor.

Constructors

GetRelationalDatabaseLogEvents'

Fields

startTime :: Maybe POSIX

The start of the time interval from which to get log events.

Constraints:

Specified in Coordinated Universal Time (UTC).

Specified in the Unix time format.

For example, if you wish to use a start time of October 1, 2018, at 8 PM UTC, then you input 1538424000 as the start time.

startFromHead :: Maybe Bool

Parameter to specify if the log should start from head or tail. If true is specified, the log event starts from the head of the log. If false is specified, the log event starts from the tail of the log.

For PostgreSQL, the default value of false is the only option available.

endTime :: Maybe POSIX

The end of the time interval from which to get log events.

Constraints:

Specified in Coordinated Universal Time (UTC).

Specified in the Unix time format.

For example, if you wish to use an end time of October 1, 2018, at 8 PM UTC, then you input 1538424000 as the end time.

pageToken :: Maybe Text

The token to advance to the next or previous page of results from your request.

To get a page token, perform an initial GetRelationalDatabaseLogEvents request. If your results are paginated, the response will return a next forward token and/or next backward token that you can specify as the page token in a subsequent request.

relationalDatabaseName :: Text

The name of your database for which to get log events.

logStreamName :: Text

The name of the log stream.

Use the get relational database log streams operation to get a list of available log streams.

newGetRelationalDatabaseLogEvents

Arguments

:: Text
$sel:relationalDatabaseName:GetRelationalDatabaseLogEvents'

-> Text
$sel:logStreamName:GetRelationalDatabaseLogEvents'

-> GetRelationalDatabaseLogEvents

Create a value of GetRelationalDatabaseLogEvents with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:startTime:GetRelationalDatabaseLogEvents', getRelationalDatabaseLogEvents_startTime - The start of the time interval from which to get log events.

Constraints:

Specified in Coordinated Universal Time (UTC).

Specified in the Unix time format.

For example, if you wish to use a start time of October 1, 2018, at 8 PM UTC, then you input 1538424000 as the start time.

$sel:startFromHead:GetRelationalDatabaseLogEvents', getRelationalDatabaseLogEvents_startFromHead - Parameter to specify if the log should start from head or tail. If true is specified, the log event starts from the head of the log. If false is specified, the log event starts from the tail of the log.

For PostgreSQL, the default value of false is the only option available.

$sel:endTime:GetRelationalDatabaseLogEvents', getRelationalDatabaseLogEvents_endTime - The end of the time interval from which to get log events.

Constraints:

Specified in Coordinated Universal Time (UTC).

Specified in the Unix time format.

For example, if you wish to use an end time of October 1, 2018, at 8 PM UTC, then you input 1538424000 as the end time.

$sel:pageToken:GetRelationalDatabaseLogEvents', getRelationalDatabaseLogEvents_pageToken - The token to advance to the next or previous page of results from your request.

To get a page token, perform an initial GetRelationalDatabaseLogEvents request. If your results are paginated, the response will return a next forward token and/or next backward token that you can specify as the page token in a subsequent request.

$sel:relationalDatabaseName:GetRelationalDatabaseLogEvents', getRelationalDatabaseLogEvents_relationalDatabaseName - The name of your database for which to get log events.

$sel:logStreamName:GetRelationalDatabaseLogEvents', getRelationalDatabaseLogEvents_logStreamName - The name of the log stream.

Use the get relational database log streams operation to get a list of available log streams.

Request Lenses

getRelationalDatabaseLogEvents_startTime :: Lens' GetRelationalDatabaseLogEvents (Maybe UTCTime)

The start of the time interval from which to get log events.

Constraints:

Specified in Coordinated Universal Time (UTC).

Specified in the Unix time format.

For example, if you wish to use a start time of October 1, 2018, at 8 PM UTC, then you input 1538424000 as the start time.

getRelationalDatabaseLogEvents_startFromHead :: Lens' GetRelationalDatabaseLogEvents (Maybe Bool)

Parameter to specify if the log should start from head or tail. If true is specified, the log event starts from the head of the log. If false is specified, the log event starts from the tail of the log.

For PostgreSQL, the default value of false is the only option available.

getRelationalDatabaseLogEvents_endTime :: Lens' GetRelationalDatabaseLogEvents (Maybe UTCTime)

The end of the time interval from which to get log events.

Constraints:

Specified in Coordinated Universal Time (UTC).

Specified in the Unix time format.

For example, if you wish to use an end time of October 1, 2018, at 8 PM UTC, then you input 1538424000 as the end time.

getRelationalDatabaseLogEvents_pageToken :: Lens' GetRelationalDatabaseLogEvents (Maybe Text)

The token to advance to the next or previous page of results from your request.

To get a page token, perform an initial GetRelationalDatabaseLogEvents request. If your results are paginated, the response will return a next forward token and/or next backward token that you can specify as the page token in a subsequent request.

getRelationalDatabaseLogEvents_relationalDatabaseName :: Lens' GetRelationalDatabaseLogEvents Text

The name of your database for which to get log events.

getRelationalDatabaseLogEvents_logStreamName :: Lens' GetRelationalDatabaseLogEvents Text

The name of the log stream.

Use the get relational database log streams operation to get a list of available log streams.

Destructuring the Response

data GetRelationalDatabaseLogEventsResponse

See: newGetRelationalDatabaseLogEventsResponse smart constructor.

Constructors

GetRelationalDatabaseLogEventsResponse'

Fields

nextBackwardToken :: Maybe Text

A token used for advancing to the previous page of results from your get relational database log events request.

resourceLogEvents :: Maybe [LogEvent]

An object describing the result of your get relational database log events request.

nextForwardToken :: Maybe Text

A token used for advancing to the next page of results from your get relational database log events request.

httpStatus :: Int

The response's http status code.

newGetRelationalDatabaseLogEventsResponse

Arguments

:: Int
$sel:httpStatus:GetRelationalDatabaseLogEventsResponse'

-> GetRelationalDatabaseLogEventsResponse

Create a value of GetRelationalDatabaseLogEventsResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:nextBackwardToken:GetRelationalDatabaseLogEventsResponse', getRelationalDatabaseLogEventsResponse_nextBackwardToken - A token used for advancing to the previous page of results from your get relational database log events request.

$sel:resourceLogEvents:GetRelationalDatabaseLogEventsResponse', getRelationalDatabaseLogEventsResponse_resourceLogEvents - An object describing the result of your get relational database log events request.

$sel:nextForwardToken:GetRelationalDatabaseLogEventsResponse', getRelationalDatabaseLogEventsResponse_nextForwardToken - A token used for advancing to the next page of results from your get relational database log events request.

$sel:httpStatus:GetRelationalDatabaseLogEventsResponse', getRelationalDatabaseLogEventsResponse_httpStatus - The response's http status code.

Response Lenses

getRelationalDatabaseLogEventsResponse_nextBackwardToken :: Lens' GetRelationalDatabaseLogEventsResponse (Maybe Text)

A token used for advancing to the previous page of results from your get relational database log events request.

getRelationalDatabaseLogEventsResponse_resourceLogEvents :: Lens' GetRelationalDatabaseLogEventsResponse (Maybe [LogEvent])

An object describing the result of your get relational database log events request.

getRelationalDatabaseLogEventsResponse_nextForwardToken :: Lens' GetRelationalDatabaseLogEventsResponse (Maybe Text)

A token used for advancing to the next page of results from your get relational database log events request.

getRelationalDatabaseLogEventsResponse_httpStatus :: Lens' GetRelationalDatabaseLogEventsResponse Int

The response's http status code.