Amazonka.Lightsail.GetRelationalDatabaseMasterUserPassword

Description

Returns the current, previous, or pending versions of the master user password for a Lightsail database.

The GetRelationalDatabaseMasterUserPassword operation supports tag-based access control via resource tags applied to the resource identified by relationalDatabaseName.

Creating a Request

data GetRelationalDatabaseMasterUserPassword

See: newGetRelationalDatabaseMasterUserPassword smart constructor.

Constructors

GetRelationalDatabaseMasterUserPassword'

Fields

passwordVersion :: Maybe RelationalDatabasePasswordVersion

The password version to return.

Specifying CURRENT or PREVIOUS returns the current or previous passwords respectively. Specifying PENDING returns the newest version of the password that will rotate to CURRENT. After the PENDING password rotates to CURRENT, the PENDING password is no longer available.

Default: CURRENT

relationalDatabaseName :: Text

The name of your database for which to get the master user password.

newGetRelationalDatabaseMasterUserPassword

Arguments

:: Text
$sel:relationalDatabaseName:GetRelationalDatabaseMasterUserPassword'

-> GetRelationalDatabaseMasterUserPassword

Create a value of GetRelationalDatabaseMasterUserPassword with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:passwordVersion:GetRelationalDatabaseMasterUserPassword', getRelationalDatabaseMasterUserPassword_passwordVersion - The password version to return.

Specifying CURRENT or PREVIOUS returns the current or previous passwords respectively. Specifying PENDING returns the newest version of the password that will rotate to CURRENT. After the PENDING password rotates to CURRENT, the PENDING password is no longer available.

Default: CURRENT

$sel:relationalDatabaseName:GetRelationalDatabaseMasterUserPassword', getRelationalDatabaseMasterUserPassword_relationalDatabaseName - The name of your database for which to get the master user password.

Request Lenses

getRelationalDatabaseMasterUserPassword_passwordVersion :: Lens' GetRelationalDatabaseMasterUserPassword (Maybe RelationalDatabasePasswordVersion)

The password version to return.

Specifying CURRENT or PREVIOUS returns the current or previous passwords respectively. Specifying PENDING returns the newest version of the password that will rotate to CURRENT. After the PENDING password rotates to CURRENT, the PENDING password is no longer available.

Default: CURRENT

getRelationalDatabaseMasterUserPassword_relationalDatabaseName :: Lens' GetRelationalDatabaseMasterUserPassword Text

The name of your database for which to get the master user password.

Destructuring the Response

data GetRelationalDatabaseMasterUserPasswordResponse

See: newGetRelationalDatabaseMasterUserPasswordResponse smart constructor.

Constructors

GetRelationalDatabaseMasterUserPasswordResponse'

Fields

masterUserPassword :: Maybe (Sensitive Text)

The master user password for the password version specified.

createdAt :: Maybe POSIX

The timestamp when the specified version of the master user password was created.

httpStatus :: Int

The response's http status code.

newGetRelationalDatabaseMasterUserPasswordResponse

Arguments

:: Int
$sel:httpStatus:GetRelationalDatabaseMasterUserPasswordResponse'

-> GetRelationalDatabaseMasterUserPasswordResponse

Create a value of GetRelationalDatabaseMasterUserPasswordResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:masterUserPassword:GetRelationalDatabaseMasterUserPasswordResponse', getRelationalDatabaseMasterUserPasswordResponse_masterUserPassword - The master user password for the password version specified.

$sel:createdAt:GetRelationalDatabaseMasterUserPasswordResponse', getRelationalDatabaseMasterUserPasswordResponse_createdAt - The timestamp when the specified version of the master user password was created.

$sel:httpStatus:GetRelationalDatabaseMasterUserPasswordResponse', getRelationalDatabaseMasterUserPasswordResponse_httpStatus - The response's http status code.

Response Lenses

getRelationalDatabaseMasterUserPasswordResponse_masterUserPassword :: Lens' GetRelationalDatabaseMasterUserPasswordResponse (Maybe Text)

The master user password for the password version specified.

getRelationalDatabaseMasterUserPasswordResponse_createdAt :: Lens' GetRelationalDatabaseMasterUserPasswordResponse (Maybe UTCTime)

The timestamp when the specified version of the master user password was created.

getRelationalDatabaseMasterUserPasswordResponse_httpStatus :: Lens' GetRelationalDatabaseMasterUserPasswordResponse Int

The response's http status code.