Amazonka.Lightsail.GetRelationalDatabaseMetricData

Description

Returns the data points of the specified metric for a database in Amazon Lightsail.

Metrics report the utilization of your resources, and the error counts generated by them. Monitor and collect metric data regularly to maintain the reliability, availability, and performance of your resources.

Creating a Request

data GetRelationalDatabaseMetricData

See: newGetRelationalDatabaseMetricData smart constructor.

Constructors

GetRelationalDatabaseMetricData'

Fields

relationalDatabaseName :: Text

The name of your database from which to get metric data.

metricName :: RelationalDatabaseMetricName

The metric for which you want to return information.

Valid relational database metric names are listed below, along with the most useful statistics to include in your request, and the published unit value. All relational database metric data is available in 1-minute (60 seconds) granularity.

CPUUtilization - The percentage of CPU utilization currently in use on the database.

Statistics: The most useful statistics are Maximum and Average.

Unit: The published unit is Percent.

DatabaseConnections - The number of database connections in use.

Statistics: The most useful statistics are Maximum and Sum.

Unit: The published unit is Count.

DiskQueueDepth - The number of outstanding IOs (read/write requests) that are waiting to access the disk.

Statistics: The most useful statistic is Sum.

Unit: The published unit is Count.

FreeStorageSpace - The amount of available storage space.

Statistics: The most useful statistic is Sum.

Unit: The published unit is Bytes.

NetworkReceiveThroughput - The incoming (Receive) network traffic on the database, including both customer database traffic and AWS traffic used for monitoring and replication.

Statistics: The most useful statistic is Average.

Unit: The published unit is Bytes/Second.

NetworkTransmitThroughput - The outgoing (Transmit) network traffic on the database, including both customer database traffic and AWS traffic used for monitoring and replication.

Statistics: The most useful statistic is Average.

Unit: The published unit is Bytes/Second.

period :: Natural

The granularity, in seconds, of the returned data points.

All relational database metric data is available in 1-minute (60 seconds) granularity.

startTime :: POSIX

The start of the time interval from which to get metric data.

Constraints:

Specified in Coordinated Universal Time (UTC).

Specified in the Unix time format.

For example, if you wish to use a start time of October 1, 2018, at 8 PM UTC, then you input 1538424000 as the start time.

endTime :: POSIX

The end of the time interval from which to get metric data.

Constraints:

Specified in Coordinated Universal Time (UTC).

Specified in the Unix time format.

For example, if you wish to use an end time of October 1, 2018, at 8 PM UTC, then you input 1538424000 as the end time.

unit :: MetricUnit

The unit for the metric data request. Valid units depend on the metric data being requested. For the valid units with each available metric, see the metricName parameter.

statistics :: [MetricStatistic]

The statistic for the metric.

The following statistics are available:

Minimum - The lowest value observed during the specified period. Use this value to determine low volumes of activity for your application.

Maximum - The highest value observed during the specified period. Use this value to determine high volumes of activity for your application.

Sum - All values submitted for the matching metric added together. You can use this statistic to determine the total volume of a metric.

Average - The value of Sum / SampleCount during the specified period. By comparing this statistic with the Minimum and Maximum values, you can determine the full scope of a metric and how close the average use is to the Minimum and Maximum values. This comparison helps you to know when to increase or decrease your resources.

SampleCount - The count, or number, of data points used for the statistical calculation.

newGetRelationalDatabaseMetricData

Arguments

:: Text
$sel:relationalDatabaseName:GetRelationalDatabaseMetricData'

-> RelationalDatabaseMetricName
$sel:metricName:GetRelationalDatabaseMetricData'

-> Natural
$sel:period:GetRelationalDatabaseMetricData'

-> UTCTime
$sel:startTime:GetRelationalDatabaseMetricData'

-> UTCTime
$sel:endTime:GetRelationalDatabaseMetricData'

-> MetricUnit
$sel:unit:GetRelationalDatabaseMetricData'

-> GetRelationalDatabaseMetricData

Create a value of GetRelationalDatabaseMetricData with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:relationalDatabaseName:GetRelationalDatabaseMetricData', getRelationalDatabaseMetricData_relationalDatabaseName - The name of your database from which to get metric data.

$sel:metricName:GetRelationalDatabaseMetricData', getRelationalDatabaseMetricData_metricName - The metric for which you want to return information.

Valid relational database metric names are listed below, along with the most useful statistics to include in your request, and the published unit value. All relational database metric data is available in 1-minute (60 seconds) granularity.

CPUUtilization - The percentage of CPU utilization currently in use on the database.

Statistics: The most useful statistics are Maximum and Average.

Unit: The published unit is Percent.

DatabaseConnections - The number of database connections in use.

Statistics: The most useful statistics are Maximum and Sum.

Unit: The published unit is Count.

DiskQueueDepth - The number of outstanding IOs (read/write requests) that are waiting to access the disk.

Statistics: The most useful statistic is Sum.

Unit: The published unit is Count.

FreeStorageSpace - The amount of available storage space.

Statistics: The most useful statistic is Sum.

Unit: The published unit is Bytes.

NetworkReceiveThroughput - The incoming (Receive) network traffic on the database, including both customer database traffic and AWS traffic used for monitoring and replication.

Statistics: The most useful statistic is Average.

Unit: The published unit is Bytes/Second.

NetworkTransmitThroughput - The outgoing (Transmit) network traffic on the database, including both customer database traffic and AWS traffic used for monitoring and replication.

Statistics: The most useful statistic is Average.

Unit: The published unit is Bytes/Second.

$sel:period:GetRelationalDatabaseMetricData', getRelationalDatabaseMetricData_period - The granularity, in seconds, of the returned data points.

All relational database metric data is available in 1-minute (60 seconds) granularity.

$sel:startTime:GetRelationalDatabaseMetricData', getRelationalDatabaseMetricData_startTime - The start of the time interval from which to get metric data.

Constraints:

Specified in Coordinated Universal Time (UTC).

Specified in the Unix time format.

For example, if you wish to use a start time of October 1, 2018, at 8 PM UTC, then you input 1538424000 as the start time.

$sel:endTime:GetRelationalDatabaseMetricData', getRelationalDatabaseMetricData_endTime - The end of the time interval from which to get metric data.

Constraints:

Specified in Coordinated Universal Time (UTC).

Specified in the Unix time format.

For example, if you wish to use an end time of October 1, 2018, at 8 PM UTC, then you input 1538424000 as the end time.

$sel:unit:GetRelationalDatabaseMetricData', getRelationalDatabaseMetricData_unit - The unit for the metric data request. Valid units depend on the metric data being requested. For the valid units with each available metric, see the metricName parameter.

$sel:statistics:GetRelationalDatabaseMetricData', getRelationalDatabaseMetricData_statistics - The statistic for the metric.

The following statistics are available:

Minimum - The lowest value observed during the specified period. Use this value to determine low volumes of activity for your application.

Maximum - The highest value observed during the specified period. Use this value to determine high volumes of activity for your application.

Sum - All values submitted for the matching metric added together. You can use this statistic to determine the total volume of a metric.

Average - The value of Sum / SampleCount during the specified period. By comparing this statistic with the Minimum and Maximum values, you can determine the full scope of a metric and how close the average use is to the Minimum and Maximum values. This comparison helps you to know when to increase or decrease your resources.

SampleCount - The count, or number, of data points used for the statistical calculation.

Request Lenses

getRelationalDatabaseMetricData_relationalDatabaseName :: Lens' GetRelationalDatabaseMetricData Text

The name of your database from which to get metric data.

getRelationalDatabaseMetricData_metricName :: Lens' GetRelationalDatabaseMetricData RelationalDatabaseMetricName

The metric for which you want to return information.

Valid relational database metric names are listed below, along with the most useful statistics to include in your request, and the published unit value. All relational database metric data is available in 1-minute (60 seconds) granularity.

CPUUtilization - The percentage of CPU utilization currently in use on the database.

Statistics: The most useful statistics are Maximum and Average.

Unit: The published unit is Percent.

DatabaseConnections - The number of database connections in use.

Statistics: The most useful statistics are Maximum and Sum.

Unit: The published unit is Count.

DiskQueueDepth - The number of outstanding IOs (read/write requests) that are waiting to access the disk.

Statistics: The most useful statistic is Sum.

Unit: The published unit is Count.

FreeStorageSpace - The amount of available storage space.

Statistics: The most useful statistic is Sum.

Unit: The published unit is Bytes.

NetworkReceiveThroughput - The incoming (Receive) network traffic on the database, including both customer database traffic and AWS traffic used for monitoring and replication.

Statistics: The most useful statistic is Average.

Unit: The published unit is Bytes/Second.

NetworkTransmitThroughput - The outgoing (Transmit) network traffic on the database, including both customer database traffic and AWS traffic used for monitoring and replication.

Statistics: The most useful statistic is Average.

Unit: The published unit is Bytes/Second.

getRelationalDatabaseMetricData_period :: Lens' GetRelationalDatabaseMetricData Natural

The granularity, in seconds, of the returned data points.

All relational database metric data is available in 1-minute (60 seconds) granularity.

getRelationalDatabaseMetricData_startTime :: Lens' GetRelationalDatabaseMetricData UTCTime

The start of the time interval from which to get metric data.

Constraints:

Specified in Coordinated Universal Time (UTC).

Specified in the Unix time format.

For example, if you wish to use a start time of October 1, 2018, at 8 PM UTC, then you input 1538424000 as the start time.

getRelationalDatabaseMetricData_endTime :: Lens' GetRelationalDatabaseMetricData UTCTime

The end of the time interval from which to get metric data.

Constraints:

Specified in Coordinated Universal Time (UTC).

Specified in the Unix time format.

For example, if you wish to use an end time of October 1, 2018, at 8 PM UTC, then you input 1538424000 as the end time.

getRelationalDatabaseMetricData_unit :: Lens' GetRelationalDatabaseMetricData MetricUnit

The unit for the metric data request. Valid units depend on the metric data being requested. For the valid units with each available metric, see the metricName parameter.

getRelationalDatabaseMetricData_statistics :: Lens' GetRelationalDatabaseMetricData [MetricStatistic]

The statistic for the metric.

The following statistics are available:

Minimum - The lowest value observed during the specified period. Use this value to determine low volumes of activity for your application.

Maximum - The highest value observed during the specified period. Use this value to determine high volumes of activity for your application.

Sum - All values submitted for the matching metric added together. You can use this statistic to determine the total volume of a metric.

Average - The value of Sum / SampleCount during the specified period. By comparing this statistic with the Minimum and Maximum values, you can determine the full scope of a metric and how close the average use is to the Minimum and Maximum values. This comparison helps you to know when to increase or decrease your resources.

SampleCount - The count, or number, of data points used for the statistical calculation.

Destructuring the Response

data GetRelationalDatabaseMetricDataResponse

See: newGetRelationalDatabaseMetricDataResponse smart constructor.

Constructors

GetRelationalDatabaseMetricDataResponse'

Fields

metricName :: Maybe RelationalDatabaseMetricName

The name of the metric returned.

metricData :: Maybe [MetricDatapoint]

An array of objects that describe the metric data returned.

httpStatus :: Int

The response's http status code.

newGetRelationalDatabaseMetricDataResponse

Arguments

:: Int
$sel:httpStatus:GetRelationalDatabaseMetricDataResponse'

-> GetRelationalDatabaseMetricDataResponse

Create a value of GetRelationalDatabaseMetricDataResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:metricName:GetRelationalDatabaseMetricData', getRelationalDatabaseMetricDataResponse_metricName - The name of the metric returned.

$sel:metricData:GetRelationalDatabaseMetricDataResponse', getRelationalDatabaseMetricDataResponse_metricData - An array of objects that describe the metric data returned.

$sel:httpStatus:GetRelationalDatabaseMetricDataResponse', getRelationalDatabaseMetricDataResponse_httpStatus - The response's http status code.

Response Lenses

getRelationalDatabaseMetricDataResponse_metricName :: Lens' GetRelationalDatabaseMetricDataResponse (Maybe RelationalDatabaseMetricName)

The name of the metric returned.

getRelationalDatabaseMetricDataResponse_metricData :: Lens' GetRelationalDatabaseMetricDataResponse (Maybe [MetricDatapoint])

An array of objects that describe the metric data returned.

getRelationalDatabaseMetricDataResponse_httpStatus :: Lens' GetRelationalDatabaseMetricDataResponse Int

The response's http status code.