Amazonka.Lightsail.RebootRelationalDatabase

Description

Restarts a specific database in Amazon Lightsail.

The reboot relational database operation supports tag-based access control via resource tags applied to the resource identified by relationalDatabaseName. For more information, see the Amazon Lightsail Developer Guide.

Creating a Request

data RebootRelationalDatabase

See: newRebootRelationalDatabase smart constructor.

Constructors

RebootRelationalDatabase'

Fields

relationalDatabaseName :: Text

The name of your database to reboot.

newRebootRelationalDatabase

Arguments

:: Text
$sel:relationalDatabaseName:RebootRelationalDatabase'

-> RebootRelationalDatabase

Create a value of RebootRelationalDatabase with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:relationalDatabaseName:RebootRelationalDatabase', rebootRelationalDatabase_relationalDatabaseName - The name of your database to reboot.

Request Lenses

rebootRelationalDatabase_relationalDatabaseName :: Lens' RebootRelationalDatabase Text

The name of your database to reboot.

Destructuring the Response

data RebootRelationalDatabaseResponse

See: newRebootRelationalDatabaseResponse smart constructor.

Constructors

RebootRelationalDatabaseResponse'

Fields

operations :: Maybe [Operation]

An array of objects that describe the result of the action, such as the status of the request, the timestamp of the request, and the resources affected by the request.

httpStatus :: Int

The response's http status code.

newRebootRelationalDatabaseResponse

Arguments

:: Int
$sel:httpStatus:RebootRelationalDatabaseResponse'

-> RebootRelationalDatabaseResponse

Create a value of RebootRelationalDatabaseResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:operations:RebootRelationalDatabaseResponse', rebootRelationalDatabaseResponse_operations - An array of objects that describe the result of the action, such as the status of the request, the timestamp of the request, and the resources affected by the request.

$sel:httpStatus:RebootRelationalDatabaseResponse', rebootRelationalDatabaseResponse_httpStatus - The response's http status code.

Response Lenses

rebootRelationalDatabaseResponse_operations :: Lens' RebootRelationalDatabaseResponse (Maybe [Operation])

An array of objects that describe the result of the action, such as the status of the request, the timestamp of the request, and the resources affected by the request.

rebootRelationalDatabaseResponse_httpStatus :: Lens' RebootRelationalDatabaseResponse Int

The response's http status code.