Amazonka.Lightsail.ReleaseStaticIp

Description

Deletes a specific static IP from your account.

Creating a Request

data ReleaseStaticIp

See: newReleaseStaticIp smart constructor.

Constructors

ReleaseStaticIp'

Fields

staticIpName :: Text

The name of the static IP to delete.

newReleaseStaticIp

Arguments

:: Text
$sel:staticIpName:ReleaseStaticIp'

-> ReleaseStaticIp

Create a value of ReleaseStaticIp with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:staticIpName:ReleaseStaticIp', releaseStaticIp_staticIpName - The name of the static IP to delete.

Request Lenses

releaseStaticIp_staticIpName :: Lens' ReleaseStaticIp Text

The name of the static IP to delete.

Destructuring the Response

data ReleaseStaticIpResponse

See: newReleaseStaticIpResponse smart constructor.

Constructors

ReleaseStaticIpResponse'

Fields

operations :: Maybe [Operation]

An array of objects that describe the result of the action, such as the status of the request, the timestamp of the request, and the resources affected by the request.

httpStatus :: Int

The response's http status code.

newReleaseStaticIpResponse

Arguments

:: Int
$sel:httpStatus:ReleaseStaticIpResponse'

-> ReleaseStaticIpResponse

Create a value of ReleaseStaticIpResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:operations:ReleaseStaticIpResponse', releaseStaticIpResponse_operations - An array of objects that describe the result of the action, such as the status of the request, the timestamp of the request, and the resources affected by the request.

$sel:httpStatus:ReleaseStaticIpResponse', releaseStaticIpResponse_httpStatus - The response's http status code.

Response Lenses

releaseStaticIpResponse_operations :: Lens' ReleaseStaticIpResponse (Maybe [Operation])

An array of objects that describe the result of the action, such as the status of the request, the timestamp of the request, and the resources affected by the request.

releaseStaticIpResponse_httpStatus :: Lens' ReleaseStaticIpResponse Int

The response's http status code.