Amazonka.Lightsail.Types.AccessRules

Description

Documentation

data AccessRules

Describes the anonymous access permissions for an Amazon Lightsail bucket and its objects.

For more information about bucket access permissions, see Understanding bucket permissions in Amazon Lightsail in the

Amazon Lightsail Developer Guide.

See: newAccessRules smart constructor.

Constructors

AccessRules'

Fields

getObject :: Maybe AccessType

Specifies the anonymous access to all objects in a bucket.

The following options can be specified:

public - Sets all objects in the bucket to public (read-only), making them readable by anyone in the world.

If the getObject value is set to public, then all objects in the bucket default to public regardless of the allowPublicOverrides value.

private - Sets all objects in the bucket to private, making them readable only by you or anyone you give access to.

If the getObject value is set to private, and the allowPublicOverrides value is set to true, then all objects in the bucket default to private unless they are configured with a public-read ACL. Individual objects with a public-read ACL are readable by anyone in the world.

allowPublicOverrides :: Maybe Bool

A Boolean value that indicates whether the access control list (ACL) permissions that are applied to individual objects override the getObject option that is currently specified.

When this is true, you can use the PutObjectAcl Amazon S3 API action to set individual objects to public (read-only) using the public-read ACL, or to private using the private ACL.

newAccessRules :: AccessRules

Create a value of AccessRules with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:getObject:AccessRules', accessRules_getObject - Specifies the anonymous access to all objects in a bucket.

The following options can be specified:

public - Sets all objects in the bucket to public (read-only), making them readable by anyone in the world.

If the getObject value is set to public, then all objects in the bucket default to public regardless of the allowPublicOverrides value.

private - Sets all objects in the bucket to private, making them readable only by you or anyone you give access to.

If the getObject value is set to private, and the allowPublicOverrides value is set to true, then all objects in the bucket default to private unless they are configured with a public-read ACL. Individual objects with a public-read ACL are readable by anyone in the world.

$sel:allowPublicOverrides:AccessRules', accessRules_allowPublicOverrides - A Boolean value that indicates whether the access control list (ACL) permissions that are applied to individual objects override the getObject option that is currently specified.

When this is true, you can use the PutObjectAcl Amazon S3 API action to set individual objects to public (read-only) using the public-read ACL, or to private using the private ACL.

accessRules_getObject :: Lens' AccessRules (Maybe AccessType)

Specifies the anonymous access to all objects in a bucket.

The following options can be specified:

public - Sets all objects in the bucket to public (read-only), making them readable by anyone in the world.

If the getObject value is set to public, then all objects in the bucket default to public regardless of the allowPublicOverrides value.

private - Sets all objects in the bucket to private, making them readable only by you or anyone you give access to.

If the getObject value is set to private, and the allowPublicOverrides value is set to true, then all objects in the bucket default to private unless they are configured with a public-read ACL. Individual objects with a public-read ACL are readable by anyone in the world.

accessRules_allowPublicOverrides :: Lens' AccessRules (Maybe Bool)

A Boolean value that indicates whether the access control list (ACL) permissions that are applied to individual objects override the getObject option that is currently specified.

When this is true, you can use the PutObjectAcl Amazon S3 API action to set individual objects to public (read-only) using the public-read ACL, or to private using the private ACL.