Amazonka.Lightsail.Types.AutoSnapshotDetails

Description

Documentation

data AutoSnapshotDetails

Describes an automatic snapshot.

See: newAutoSnapshotDetails smart constructor.

Constructors

AutoSnapshotDetails'

Fields

status :: Maybe AutoSnapshotStatus

The status of the automatic snapshot.

fromAttachedDisks :: Maybe [AttachedDisk]

An array of objects that describe the block storage disks attached to the instance when the automatic snapshot was created.

createdAt :: Maybe POSIX

The timestamp when the automatic snapshot was created.

date :: Maybe Text

The date of the automatic snapshot in YYYY-MM-DD format.

newAutoSnapshotDetails :: AutoSnapshotDetails

Create a value of AutoSnapshotDetails with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:status:AutoSnapshotDetails', autoSnapshotDetails_status - The status of the automatic snapshot.

$sel:fromAttachedDisks:AutoSnapshotDetails', autoSnapshotDetails_fromAttachedDisks - An array of objects that describe the block storage disks attached to the instance when the automatic snapshot was created.

$sel:createdAt:AutoSnapshotDetails', autoSnapshotDetails_createdAt - The timestamp when the automatic snapshot was created.

$sel:date:AutoSnapshotDetails', autoSnapshotDetails_date - The date of the automatic snapshot in YYYY-MM-DD format.

autoSnapshotDetails_status :: Lens' AutoSnapshotDetails (Maybe AutoSnapshotStatus)

The status of the automatic snapshot.

autoSnapshotDetails_fromAttachedDisks :: Lens' AutoSnapshotDetails (Maybe [AttachedDisk])

An array of objects that describe the block storage disks attached to the instance when the automatic snapshot was created.

autoSnapshotDetails_createdAt :: Lens' AutoSnapshotDetails (Maybe UTCTime)

The timestamp when the automatic snapshot was created.

autoSnapshotDetails_date :: Lens' AutoSnapshotDetails (Maybe Text)

The date of the automatic snapshot in YYYY-MM-DD format.