Amazonka.Lightsail.Types.Bucket

Description

Documentation

data Bucket

Describes an Amazon Lightsail bucket.

See: newBucket smart constructor.

Constructors

Bucket'

Fields

state :: Maybe BucketState

An object that describes the state of the bucket.

objectVersioning :: Maybe Text

Indicates whether object versioning is enabled for the bucket.

The following options can be configured:

Enabled - Object versioning is enabled.

Suspended - Object versioning was previously enabled but is currently suspended. Existing object versions are retained.

NeverEnabled - Object versioning has never been enabled.

resourceType :: Maybe Text

The Lightsail resource type of the bucket (for example, Bucket).

arn :: Maybe Text

The Amazon Resource Name (ARN) of the bucket.

createdAt :: Maybe POSIX

The timestamp when the distribution was created.

location :: Maybe ResourceLocation

readonlyAccessAccounts :: Maybe [Text]

An array of strings that specify the AWS account IDs that have read-only access to the bucket.

url :: Maybe Text

The URL of the bucket.

resourcesReceivingAccess :: Maybe [ResourceReceivingAccess]

An array of objects that describe Lightsail instances that have access to the bucket.

Use the SetResourceAccessForBucket action to update the instances that have access to a bucket.

bundleId :: Maybe Text

The ID of the bundle currently applied to the bucket.

A bucket bundle specifies the monthly cost, storage space, and data transfer quota for a bucket.

Use the UpdateBucketBundle action to change the bundle of a bucket.

accessRules :: Maybe AccessRules

An object that describes the access rules of the bucket.

ableToUpdateBundle :: Maybe Bool

Indicates whether the bundle that is currently applied to a bucket can be changed to another bundle.

You can update a bucket's bundle only one time within a monthly AWS billing cycle.

Use the UpdateBucketBundle action to change a bucket's bundle.

name :: Maybe Text

The name of the bucket.

supportCode :: Maybe Text

The support code for a bucket. Include this code in your email to support when you have questions about a Lightsail bucket. This code enables our support team to look up your Lightsail information more easily.

tags :: Maybe [Tag]

The tag keys and optional values for the bucket. For more information, see Tags in Amazon Lightsail in the Amazon Lightsail Developer Guide.

newBucket :: Bucket

Create a value of Bucket with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:state:Bucket', bucket_state - An object that describes the state of the bucket.

$sel:objectVersioning:Bucket', bucket_objectVersioning - Indicates whether object versioning is enabled for the bucket.

The following options can be configured:

Enabled - Object versioning is enabled.

Suspended - Object versioning was previously enabled but is currently suspended. Existing object versions are retained.

NeverEnabled - Object versioning has never been enabled.

$sel:resourceType:Bucket', bucket_resourceType - The Lightsail resource type of the bucket (for example, Bucket).

$sel:arn:Bucket', bucket_arn - The Amazon Resource Name (ARN) of the bucket.

$sel:createdAt:Bucket', bucket_createdAt - The timestamp when the distribution was created.

$sel:location:Bucket', bucket_location - Undocumented member.

$sel:readonlyAccessAccounts:Bucket', bucket_readonlyAccessAccounts - An array of strings that specify the AWS account IDs that have read-only access to the bucket.

$sel:url:Bucket', bucket_url - The URL of the bucket.

$sel:resourcesReceivingAccess:Bucket', bucket_resourcesReceivingAccess - An array of objects that describe Lightsail instances that have access to the bucket.

Use the SetResourceAccessForBucket action to update the instances that have access to a bucket.

$sel:bundleId:Bucket', bucket_bundleId - The ID of the bundle currently applied to the bucket.

A bucket bundle specifies the monthly cost, storage space, and data transfer quota for a bucket.

Use the UpdateBucketBundle action to change the bundle of a bucket.

$sel:accessRules:Bucket', bucket_accessRules - An object that describes the access rules of the bucket.

$sel:ableToUpdateBundle:Bucket', bucket_ableToUpdateBundle - Indicates whether the bundle that is currently applied to a bucket can be changed to another bundle.

You can update a bucket's bundle only one time within a monthly AWS billing cycle.

Use the UpdateBucketBundle action to change a bucket's bundle.

$sel:name:Bucket', bucket_name - The name of the bucket.

$sel:supportCode:Bucket', bucket_supportCode - The support code for a bucket. Include this code in your email to support when you have questions about a Lightsail bucket. This code enables our support team to look up your Lightsail information more easily.

$sel:tags:Bucket', bucket_tags - The tag keys and optional values for the bucket. For more information, see Tags in Amazon Lightsail in the Amazon Lightsail Developer Guide.

bucket_state :: Lens' Bucket (Maybe BucketState)

An object that describes the state of the bucket.

bucket_objectVersioning :: Lens' Bucket (Maybe Text)

Indicates whether object versioning is enabled for the bucket.

The following options can be configured:

Enabled - Object versioning is enabled.

Suspended - Object versioning was previously enabled but is currently suspended. Existing object versions are retained.

NeverEnabled - Object versioning has never been enabled.

bucket_resourceType :: Lens' Bucket (Maybe Text)

The Lightsail resource type of the bucket (for example, Bucket).

bucket_arn :: Lens' Bucket (Maybe Text)

The Amazon Resource Name (ARN) of the bucket.

bucket_createdAt :: Lens' Bucket (Maybe UTCTime)

The timestamp when the distribution was created.

bucket_location :: Lens' Bucket (Maybe ResourceLocation)

Undocumented member.

bucket_readonlyAccessAccounts :: Lens' Bucket (Maybe [Text])

An array of strings that specify the AWS account IDs that have read-only access to the bucket.

bucket_url :: Lens' Bucket (Maybe Text)

The URL of the bucket.

bucket_resourcesReceivingAccess :: Lens' Bucket (Maybe [ResourceReceivingAccess])

An array of objects that describe Lightsail instances that have access to the bucket.

Use the SetResourceAccessForBucket action to update the instances that have access to a bucket.

bucket_bundleId :: Lens' Bucket (Maybe Text)

The ID of the bundle currently applied to the bucket.

A bucket bundle specifies the monthly cost, storage space, and data transfer quota for a bucket.

Use the UpdateBucketBundle action to change the bundle of a bucket.

bucket_accessRules :: Lens' Bucket (Maybe AccessRules)

An object that describes the access rules of the bucket.

bucket_ableToUpdateBundle :: Lens' Bucket (Maybe Bool)

Indicates whether the bundle that is currently applied to a bucket can be changed to another bundle.

You can update a bucket's bundle only one time within a monthly AWS billing cycle.

Use the UpdateBucketBundle action to change a bucket's bundle.

bucket_name :: Lens' Bucket (Maybe Text)

The name of the bucket.

bucket_supportCode :: Lens' Bucket (Maybe Text)

The support code for a bucket. Include this code in your email to support when you have questions about a Lightsail bucket. This code enables our support team to look up your Lightsail information more easily.

bucket_tags :: Lens' Bucket (Maybe [Tag])

The tag keys and optional values for the bucket. For more information, see Tags in Amazon Lightsail in the Amazon Lightsail Developer Guide.