Amazonka.Lightsail.Types.ContactMethod

Description

Documentation

data ContactMethod

Describes a contact method.

A contact method is a way to send you notifications. For more information, see Notifications in Amazon Lightsail.

See: newContactMethod smart constructor.

Constructors

ContactMethod'

Fields

status :: Maybe ContactMethodStatus

The current status of the contact method.

A contact method has the following possible status:

PendingVerification - The contact method has not yet been verified, and the verification has not yet expired.

Valid - The contact method has been verified.

InValid - An attempt was made to verify the contact method, but the verification has expired.

resourceType :: Maybe ResourceType

The Lightsail resource type (e.g., ContactMethod).

arn :: Maybe Text

The Amazon Resource Name (ARN) of the contact method.

createdAt :: Maybe POSIX

The timestamp when the contact method was created.

location :: Maybe ResourceLocation

protocol :: Maybe ContactProtocol

The protocol of the contact method, such as email or SMS (text messaging).

name :: Maybe Text

The name of the contact method.

supportCode :: Maybe Text

The support code. Include this code in your email to support when you have questions about your Lightsail contact method. This code enables our support team to look up your Lightsail information more easily.

contactEndpoint :: Maybe Text

The destination of the contact method, such as an email address or a mobile phone number.

newContactMethod :: ContactMethod

Create a value of ContactMethod with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:status:ContactMethod', contactMethod_status - The current status of the contact method.

A contact method has the following possible status:

PendingVerification - The contact method has not yet been verified, and the verification has not yet expired.

Valid - The contact method has been verified.

InValid - An attempt was made to verify the contact method, but the verification has expired.

$sel:resourceType:ContactMethod', contactMethod_resourceType - The Lightsail resource type (e.g., ContactMethod).

$sel:arn:ContactMethod', contactMethod_arn - The Amazon Resource Name (ARN) of the contact method.

$sel:createdAt:ContactMethod', contactMethod_createdAt - The timestamp when the contact method was created.

$sel:location:ContactMethod', contactMethod_location - Undocumented member.

$sel:protocol:ContactMethod', contactMethod_protocol - The protocol of the contact method, such as email or SMS (text messaging).

$sel:name:ContactMethod', contactMethod_name - The name of the contact method.

$sel:supportCode:ContactMethod', contactMethod_supportCode - The support code. Include this code in your email to support when you have questions about your Lightsail contact method. This code enables our support team to look up your Lightsail information more easily.

$sel:contactEndpoint:ContactMethod', contactMethod_contactEndpoint - The destination of the contact method, such as an email address or a mobile phone number.

contactMethod_status :: Lens' ContactMethod (Maybe ContactMethodStatus)

The current status of the contact method.

A contact method has the following possible status:

PendingVerification - The contact method has not yet been verified, and the verification has not yet expired.

Valid - The contact method has been verified.

InValid - An attempt was made to verify the contact method, but the verification has expired.

contactMethod_resourceType :: Lens' ContactMethod (Maybe ResourceType)

The Lightsail resource type (e.g., ContactMethod).

contactMethod_arn :: Lens' ContactMethod (Maybe Text)

The Amazon Resource Name (ARN) of the contact method.

contactMethod_createdAt :: Lens' ContactMethod (Maybe UTCTime)

The timestamp when the contact method was created.

contactMethod_location :: Lens' ContactMethod (Maybe ResourceLocation)

Undocumented member.

contactMethod_protocol :: Lens' ContactMethod (Maybe ContactProtocol)

The protocol of the contact method, such as email or SMS (text messaging).

contactMethod_name :: Lens' ContactMethod (Maybe Text)

The name of the contact method.

contactMethod_supportCode :: Lens' ContactMethod (Maybe Text)

The support code. Include this code in your email to support when you have questions about your Lightsail contact method. This code enables our support team to look up your Lightsail information more easily.

contactMethod_contactEndpoint :: Lens' ContactMethod (Maybe Text)

The destination of the contact method, such as an email address or a mobile phone number.