Amazonka.Lightsail.Types.DiskSnapshot

Description

Documentation

data DiskSnapshot

Describes a block storage disk snapshot.

See: newDiskSnapshot smart constructor.

Constructors

DiskSnapshot'

Fields

fromDiskName :: Maybe Text

The unique name of the source disk from which the disk snapshot was created.

isFromAutoSnapshot :: Maybe Bool

A Boolean value indicating whether the snapshot was created from an automatic snapshot.

state :: Maybe DiskSnapshotState

The status of the disk snapshot operation.

resourceType :: Maybe ResourceType

The Lightsail resource type (e.g., DiskSnapshot).

arn :: Maybe Text

The Amazon Resource Name (ARN) of the disk snapshot.

createdAt :: Maybe POSIX

The date when the disk snapshot was created.

location :: Maybe ResourceLocation

The AWS Region and Availability Zone where the disk snapshot was created.

progress :: Maybe Text

The progress of the snapshot.

name :: Maybe Text

The name of the disk snapshot (e.g., my-disk-snapshot).

sizeInGb :: Maybe Int

The size of the disk in GB.

supportCode :: Maybe Text

The support code. Include this code in your email to support when you have questions about an instance or another resource in Lightsail. This code enables our support team to look up your Lightsail information more easily.

fromInstanceArn :: Maybe Text

The Amazon Resource Name (ARN) of the source instance from which the disk (system volume) snapshot was created.

fromInstanceName :: Maybe Text

The unique name of the source instance from which the disk (system volume) snapshot was created.

fromDiskArn :: Maybe Text

The Amazon Resource Name (ARN) of the source disk from which the disk snapshot was created.

tags :: Maybe [Tag]

The tag keys and optional values for the resource. For more information about tags in Lightsail, see the Amazon Lightsail Developer Guide.

newDiskSnapshot :: DiskSnapshot

Create a value of DiskSnapshot with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:fromDiskName:DiskSnapshot', diskSnapshot_fromDiskName - The unique name of the source disk from which the disk snapshot was created.

$sel:isFromAutoSnapshot:DiskSnapshot', diskSnapshot_isFromAutoSnapshot - A Boolean value indicating whether the snapshot was created from an automatic snapshot.

$sel:state:DiskSnapshot', diskSnapshot_state - The status of the disk snapshot operation.

$sel:resourceType:DiskSnapshot', diskSnapshot_resourceType - The Lightsail resource type (e.g., DiskSnapshot).

$sel:arn:DiskSnapshot', diskSnapshot_arn - The Amazon Resource Name (ARN) of the disk snapshot.

$sel:createdAt:DiskSnapshot', diskSnapshot_createdAt - The date when the disk snapshot was created.

$sel:location:DiskSnapshot', diskSnapshot_location - The AWS Region and Availability Zone where the disk snapshot was created.

$sel:progress:DiskSnapshot', diskSnapshot_progress - The progress of the snapshot.

$sel:name:DiskSnapshot', diskSnapshot_name - The name of the disk snapshot (e.g., my-disk-snapshot).

$sel:sizeInGb:DiskSnapshot', diskSnapshot_sizeInGb - The size of the disk in GB.

$sel:supportCode:DiskSnapshot', diskSnapshot_supportCode - The support code. Include this code in your email to support when you have questions about an instance or another resource in Lightsail. This code enables our support team to look up your Lightsail information more easily.

$sel:fromInstanceArn:DiskSnapshot', diskSnapshot_fromInstanceArn - The Amazon Resource Name (ARN) of the source instance from which the disk (system volume) snapshot was created.

$sel:fromInstanceName:DiskSnapshot', diskSnapshot_fromInstanceName - The unique name of the source instance from which the disk (system volume) snapshot was created.

$sel:fromDiskArn:DiskSnapshot', diskSnapshot_fromDiskArn - The Amazon Resource Name (ARN) of the source disk from which the disk snapshot was created.

$sel:tags:DiskSnapshot', diskSnapshot_tags - The tag keys and optional values for the resource. For more information about tags in Lightsail, see the Amazon Lightsail Developer Guide.

diskSnapshot_fromDiskName :: Lens' DiskSnapshot (Maybe Text)

The unique name of the source disk from which the disk snapshot was created.

diskSnapshot_isFromAutoSnapshot :: Lens' DiskSnapshot (Maybe Bool)

A Boolean value indicating whether the snapshot was created from an automatic snapshot.

diskSnapshot_state :: Lens' DiskSnapshot (Maybe DiskSnapshotState)

The status of the disk snapshot operation.

diskSnapshot_resourceType :: Lens' DiskSnapshot (Maybe ResourceType)

The Lightsail resource type (e.g., DiskSnapshot).

diskSnapshot_arn :: Lens' DiskSnapshot (Maybe Text)

The Amazon Resource Name (ARN) of the disk snapshot.

diskSnapshot_createdAt :: Lens' DiskSnapshot (Maybe UTCTime)

The date when the disk snapshot was created.

diskSnapshot_location :: Lens' DiskSnapshot (Maybe ResourceLocation)

The AWS Region and Availability Zone where the disk snapshot was created.

diskSnapshot_progress :: Lens' DiskSnapshot (Maybe Text)

The progress of the snapshot.

diskSnapshot_name :: Lens' DiskSnapshot (Maybe Text)

The name of the disk snapshot (e.g., my-disk-snapshot).

diskSnapshot_sizeInGb :: Lens' DiskSnapshot (Maybe Int)

The size of the disk in GB.

diskSnapshot_supportCode :: Lens' DiskSnapshot (Maybe Text)

The support code. Include this code in your email to support when you have questions about an instance or another resource in Lightsail. This code enables our support team to look up your Lightsail information more easily.

diskSnapshot_fromInstanceArn :: Lens' DiskSnapshot (Maybe Text)

The Amazon Resource Name (ARN) of the source instance from which the disk (system volume) snapshot was created.

diskSnapshot_fromInstanceName :: Lens' DiskSnapshot (Maybe Text)

The unique name of the source instance from which the disk (system volume) snapshot was created.

diskSnapshot_fromDiskArn :: Lens' DiskSnapshot (Maybe Text)

The Amazon Resource Name (ARN) of the source disk from which the disk snapshot was created.

diskSnapshot_tags :: Lens' DiskSnapshot (Maybe [Tag])

The tag keys and optional values for the resource. For more information about tags in Lightsail, see the Amazon Lightsail Developer Guide.