Amazonka.Lightsail.Types.ExportSnapshotRecord

Description

Documentation

data ExportSnapshotRecord

Describes an export snapshot record.

See: newExportSnapshotRecord smart constructor.

Constructors

ExportSnapshotRecord'

Fields

state :: Maybe RecordState

The state of the export snapshot record.

destinationInfo :: Maybe DestinationInfo

A list of objects describing the destination of the export snapshot record.

resourceType :: Maybe ResourceType

The Lightsail resource type (e.g., ExportSnapshotRecord).

arn :: Maybe Text

The Amazon Resource Name (ARN) of the export snapshot record.

createdAt :: Maybe POSIX

The date when the export snapshot record was created.

location :: Maybe ResourceLocation

The AWS Region and Availability Zone where the export snapshot record is located.

name :: Maybe Text

The export snapshot record name.

sourceInfo :: Maybe ExportSnapshotRecordSourceInfo

A list of objects describing the source of the export snapshot record.

newExportSnapshotRecord :: ExportSnapshotRecord

Create a value of ExportSnapshotRecord with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:state:ExportSnapshotRecord', exportSnapshotRecord_state - The state of the export snapshot record.

$sel:destinationInfo:ExportSnapshotRecord', exportSnapshotRecord_destinationInfo - A list of objects describing the destination of the export snapshot record.

$sel:resourceType:ExportSnapshotRecord', exportSnapshotRecord_resourceType - The Lightsail resource type (e.g., ExportSnapshotRecord).

$sel:arn:ExportSnapshotRecord', exportSnapshotRecord_arn - The Amazon Resource Name (ARN) of the export snapshot record.

$sel:createdAt:ExportSnapshotRecord', exportSnapshotRecord_createdAt - The date when the export snapshot record was created.

$sel:location:ExportSnapshotRecord', exportSnapshotRecord_location - The AWS Region and Availability Zone where the export snapshot record is located.

$sel:name:ExportSnapshotRecord', exportSnapshotRecord_name - The export snapshot record name.

$sel:sourceInfo:ExportSnapshotRecord', exportSnapshotRecord_sourceInfo - A list of objects describing the source of the export snapshot record.

exportSnapshotRecord_state :: Lens' ExportSnapshotRecord (Maybe RecordState)

The state of the export snapshot record.

exportSnapshotRecord_destinationInfo :: Lens' ExportSnapshotRecord (Maybe DestinationInfo)

A list of objects describing the destination of the export snapshot record.

exportSnapshotRecord_resourceType :: Lens' ExportSnapshotRecord (Maybe ResourceType)

The Lightsail resource type (e.g., ExportSnapshotRecord).

exportSnapshotRecord_arn :: Lens' ExportSnapshotRecord (Maybe Text)

The Amazon Resource Name (ARN) of the export snapshot record.

exportSnapshotRecord_createdAt :: Lens' ExportSnapshotRecord (Maybe UTCTime)

The date when the export snapshot record was created.

exportSnapshotRecord_location :: Lens' ExportSnapshotRecord (Maybe ResourceLocation)

The AWS Region and Availability Zone where the export snapshot record is located.

exportSnapshotRecord_name :: Lens' ExportSnapshotRecord (Maybe Text)

The export snapshot record name.

exportSnapshotRecord_sourceInfo :: Lens' ExportSnapshotRecord (Maybe ExportSnapshotRecordSourceInfo)

A list of objects describing the source of the export snapshot record.