Amazonka.Lightsail.Types.ExportSnapshotRecordSourceInfo

Description

Documentation

data ExportSnapshotRecordSourceInfo

Describes the source of an export snapshot record.

See: newExportSnapshotRecordSourceInfo smart constructor.

Constructors

ExportSnapshotRecordSourceInfo'

Fields

diskSnapshotInfo :: Maybe DiskSnapshotInfo

A list of objects describing a disk snapshot.

resourceType :: Maybe ExportSnapshotRecordSourceType

The Lightsail resource type (e.g., InstanceSnapshot or DiskSnapshot).

arn :: Maybe Text

The Amazon Resource Name (ARN) of the source instance or disk snapshot.

createdAt :: Maybe POSIX

The date when the source instance or disk snapshot was created.

fromResourceArn :: Maybe Text

The Amazon Resource Name (ARN) of the snapshot's source instance or disk.

name :: Maybe Text

The name of the source instance or disk snapshot.

instanceSnapshotInfo :: Maybe InstanceSnapshotInfo

A list of objects describing an instance snapshot.

fromResourceName :: Maybe Text

The name of the snapshot's source instance or disk.

newExportSnapshotRecordSourceInfo :: ExportSnapshotRecordSourceInfo

Create a value of ExportSnapshotRecordSourceInfo with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:diskSnapshotInfo:ExportSnapshotRecordSourceInfo', exportSnapshotRecordSourceInfo_diskSnapshotInfo - A list of objects describing a disk snapshot.

$sel:resourceType:ExportSnapshotRecordSourceInfo', exportSnapshotRecordSourceInfo_resourceType - The Lightsail resource type (e.g., InstanceSnapshot or DiskSnapshot).

$sel:arn:ExportSnapshotRecordSourceInfo', exportSnapshotRecordSourceInfo_arn - The Amazon Resource Name (ARN) of the source instance or disk snapshot.

$sel:createdAt:ExportSnapshotRecordSourceInfo', exportSnapshotRecordSourceInfo_createdAt - The date when the source instance or disk snapshot was created.

$sel:fromResourceArn:ExportSnapshotRecordSourceInfo', exportSnapshotRecordSourceInfo_fromResourceArn - The Amazon Resource Name (ARN) of the snapshot's source instance or disk.

$sel:name:ExportSnapshotRecordSourceInfo', exportSnapshotRecordSourceInfo_name - The name of the source instance or disk snapshot.

$sel:instanceSnapshotInfo:ExportSnapshotRecordSourceInfo', exportSnapshotRecordSourceInfo_instanceSnapshotInfo - A list of objects describing an instance snapshot.

$sel:fromResourceName:ExportSnapshotRecordSourceInfo', exportSnapshotRecordSourceInfo_fromResourceName - The name of the snapshot's source instance or disk.

exportSnapshotRecordSourceInfo_diskSnapshotInfo :: Lens' ExportSnapshotRecordSourceInfo (Maybe DiskSnapshotInfo)

A list of objects describing a disk snapshot.

exportSnapshotRecordSourceInfo_resourceType :: Lens' ExportSnapshotRecordSourceInfo (Maybe ExportSnapshotRecordSourceType)

The Lightsail resource type (e.g., InstanceSnapshot or DiskSnapshot).

exportSnapshotRecordSourceInfo_arn :: Lens' ExportSnapshotRecordSourceInfo (Maybe Text)

The Amazon Resource Name (ARN) of the source instance or disk snapshot.

exportSnapshotRecordSourceInfo_createdAt :: Lens' ExportSnapshotRecordSourceInfo (Maybe UTCTime)

The date when the source instance or disk snapshot was created.

exportSnapshotRecordSourceInfo_fromResourceArn :: Lens' ExportSnapshotRecordSourceInfo (Maybe Text)

The Amazon Resource Name (ARN) of the snapshot's source instance or disk.

exportSnapshotRecordSourceInfo_name :: Lens' ExportSnapshotRecordSourceInfo (Maybe Text)

The name of the source instance or disk snapshot.

exportSnapshotRecordSourceInfo_instanceSnapshotInfo :: Lens' ExportSnapshotRecordSourceInfo (Maybe InstanceSnapshotInfo)

A list of objects describing an instance snapshot.

exportSnapshotRecordSourceInfo_fromResourceName :: Lens' ExportSnapshotRecordSourceInfo (Maybe Text)

The name of the snapshot's source instance or disk.