Amazonka.Lightsail.Types.Instance

Description

Documentation

data Instance

Describes an instance (a virtual private server).

See: newInstance smart constructor.

Constructors

Instance'

Fields

state :: Maybe InstanceState

The status code and the state (e.g., running) for the instance.

resourceType :: Maybe ResourceType

The type of resource (usually Instance).

arn :: Maybe Text

The Amazon Resource Name (ARN) of the instance (e.g., arn:aws:lightsail:us-east-2:123456789101:Instance/244ad76f-8aad-4741-809f-12345EXAMPLE).

createdAt :: Maybe POSIX

The timestamp when the instance was created (e.g., 1479734909.17) in Unix time format.

location :: Maybe ResourceLocation

The region name and Availability Zone where the instance is located.

sshKeyName :: Maybe Text

The name of the SSH key being used to connect to the instance (e.g., LightsailDefaultKeyPair).

addOns :: Maybe [AddOn]

An array of objects representing the add-ons enabled on the instance.

username :: Maybe Text

The user name for connecting to the instance (e.g., ec2-user).

networking :: Maybe InstanceNetworking

Information about the public ports and monthly data transfer rates for the instance.

bundleId :: Maybe Text

The bundle for the instance (e.g., micro_1_0).

name :: Maybe Text

The name the user gave the instance (e.g., Amazon_Linux-1GB-Ohio-1).

ipAddressType :: Maybe IpAddressType

The IP address type of the instance.

The possible values are ipv4 for IPv4 only, and dualstack for IPv4 and IPv6.

supportCode :: Maybe Text

The support code. Include this code in your email to support when you have questions about an instance or another resource in Lightsail. This code enables our support team to look up your Lightsail information more easily.

blueprintId :: Maybe Text

The blueprint ID (e.g., os_amlinux_2016_03).

privateIpAddress :: Maybe Text

The private IP address of the instance.

blueprintName :: Maybe Text

The friendly name of the blueprint (e.g., Amazon Linux).

isStaticIp :: Maybe Bool

A Boolean value indicating whether this instance has a static IP assigned to it.

publicIpAddress :: Maybe Text

The public IP address of the instance.

hardware :: Maybe InstanceHardware

The size of the vCPU and the amount of RAM for the instance.

ipv6Addresses :: Maybe [Text]

The IPv6 addresses of the instance.

tags :: Maybe [Tag]

The tag keys and optional values for the resource. For more information about tags in Lightsail, see the Amazon Lightsail Developer Guide.

newInstance :: Instance

Create a value of Instance with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:state:Instance', instance_state - The status code and the state (e.g., running) for the instance.

$sel:resourceType:Instance', instance_resourceType - The type of resource (usually Instance).

$sel:arn:Instance', instance_arn - The Amazon Resource Name (ARN) of the instance (e.g., arn:aws:lightsail:us-east-2:123456789101:Instance/244ad76f-8aad-4741-809f-12345EXAMPLE).

$sel:createdAt:Instance', instance_createdAt - The timestamp when the instance was created (e.g., 1479734909.17) in Unix time format.

$sel:location:Instance', instance_location - The region name and Availability Zone where the instance is located.

$sel:sshKeyName:Instance', instance_sshKeyName - The name of the SSH key being used to connect to the instance (e.g., LightsailDefaultKeyPair).

$sel:addOns:Instance', instance_addOns - An array of objects representing the add-ons enabled on the instance.

$sel:username:Instance', instance_username - The user name for connecting to the instance (e.g., ec2-user).

$sel:networking:Instance', instance_networking - Information about the public ports and monthly data transfer rates for the instance.

$sel:bundleId:Instance', instance_bundleId - The bundle for the instance (e.g., micro_1_0).

$sel:name:Instance', instance_name - The name the user gave the instance (e.g., Amazon_Linux-1GB-Ohio-1).

$sel:ipAddressType:Instance', instance_ipAddressType - The IP address type of the instance.

The possible values are ipv4 for IPv4 only, and dualstack for IPv4 and IPv6.

$sel:supportCode:Instance', instance_supportCode - The support code. Include this code in your email to support when you have questions about an instance or another resource in Lightsail. This code enables our support team to look up your Lightsail information more easily.

$sel:blueprintId:Instance', instance_blueprintId - The blueprint ID (e.g., os_amlinux_2016_03).

$sel:privateIpAddress:Instance', instance_privateIpAddress - The private IP address of the instance.

$sel:blueprintName:Instance', instance_blueprintName - The friendly name of the blueprint (e.g., Amazon Linux).

$sel:isStaticIp:Instance', instance_isStaticIp - A Boolean value indicating whether this instance has a static IP assigned to it.

$sel:publicIpAddress:Instance', instance_publicIpAddress - The public IP address of the instance.

$sel:hardware:Instance', instance_hardware - The size of the vCPU and the amount of RAM for the instance.

$sel:ipv6Addresses:Instance', instance_ipv6Addresses - The IPv6 addresses of the instance.

$sel:tags:Instance', instance_tags - The tag keys and optional values for the resource. For more information about tags in Lightsail, see the Amazon Lightsail Developer Guide.

instance_state :: Lens' Instance (Maybe InstanceState)

The status code and the state (e.g., running) for the instance.

instance_resourceType :: Lens' Instance (Maybe ResourceType)

The type of resource (usually Instance).

instance_arn :: Lens' Instance (Maybe Text)

The Amazon Resource Name (ARN) of the instance (e.g., arn:aws:lightsail:us-east-2:123456789101:Instance/244ad76f-8aad-4741-809f-12345EXAMPLE).

instance_createdAt :: Lens' Instance (Maybe UTCTime)

The timestamp when the instance was created (e.g., 1479734909.17) in Unix time format.

instance_location :: Lens' Instance (Maybe ResourceLocation)

The region name and Availability Zone where the instance is located.

instance_sshKeyName :: Lens' Instance (Maybe Text)

The name of the SSH key being used to connect to the instance (e.g., LightsailDefaultKeyPair).

instance_addOns :: Lens' Instance (Maybe [AddOn])

An array of objects representing the add-ons enabled on the instance.

instance_username :: Lens' Instance (Maybe Text)

The user name for connecting to the instance (e.g., ec2-user).

instance_networking :: Lens' Instance (Maybe InstanceNetworking)

Information about the public ports and monthly data transfer rates for the instance.

instance_bundleId :: Lens' Instance (Maybe Text)

The bundle for the instance (e.g., micro_1_0).

instance_name :: Lens' Instance (Maybe Text)

The name the user gave the instance (e.g., Amazon_Linux-1GB-Ohio-1).

instance_ipAddressType :: Lens' Instance (Maybe IpAddressType)

The IP address type of the instance.

The possible values are ipv4 for IPv4 only, and dualstack for IPv4 and IPv6.

instance_supportCode :: Lens' Instance (Maybe Text)

The support code. Include this code in your email to support when you have questions about an instance or another resource in Lightsail. This code enables our support team to look up your Lightsail information more easily.

instance_blueprintId :: Lens' Instance (Maybe Text)

The blueprint ID (e.g., os_amlinux_2016_03).

instance_privateIpAddress :: Lens' Instance (Maybe Text)

The private IP address of the instance.

instance_blueprintName :: Lens' Instance (Maybe Text)

The friendly name of the blueprint (e.g., Amazon Linux).

instance_isStaticIp :: Lens' Instance (Maybe Bool)

A Boolean value indicating whether this instance has a static IP assigned to it.

instance_publicIpAddress :: Lens' Instance (Maybe Text)

The public IP address of the instance.

instance_hardware :: Lens' Instance (Maybe InstanceHardware)

The size of the vCPU and the amount of RAM for the instance.

instance_ipv6Addresses :: Lens' Instance (Maybe [Text])

The IPv6 addresses of the instance.

instance_tags :: Lens' Instance (Maybe [Tag])

The tag keys and optional values for the resource. For more information about tags in Lightsail, see the Amazon Lightsail Developer Guide.