Amazonka.Lightsail.Types.InstanceHealthSummary

Description

Documentation

data InstanceHealthSummary

Describes information about the health of the instance.

See: newInstanceHealthSummary smart constructor.

Constructors

InstanceHealthSummary'

Fields

instanceHealth :: Maybe InstanceHealthState

Describes the overall instance health. Valid values are below.

instanceName :: Maybe Text

The name of the Lightsail instance for which you are requesting health check data.

instanceHealthReason :: Maybe InstanceHealthReason

More information about the instance health. If the instanceHealth is healthy, then an instanceHealthReason value is not provided.

If instanceHealth is initial, the instanceHealthReason value can be one of the following:

Lb.RegistrationInProgress - The target instance is in the process of being registered with the load balancer.

Lb.InitialHealthChecking - The Lightsail load balancer is still sending the target instance the minimum number of health checks required to determine its health status.

If instanceHealth is unhealthy, the instanceHealthReason value can be one of the following:

Instance.ResponseCodeMismatch - The health checks did not return an expected HTTP code.

Instance.Timeout - The health check requests timed out.

Instance.FailedHealthChecks - The health checks failed because the connection to the target instance timed out, the target instance response was malformed, or the target instance failed the health check for an unknown reason.

Lb.InternalError - The health checks failed due to an internal error.

If instanceHealth is unused, the instanceHealthReason value can be one of the following:

Instance.NotRegistered - The target instance is not registered with the target group.

Instance.NotInUse - The target group is not used by any load balancer, or the target instance is in an Availability Zone that is not enabled for its load balancer.

Instance.IpUnusable - The target IP address is reserved for use by a Lightsail load balancer.

Instance.InvalidState - The target is in the stopped or terminated state.

If instanceHealth is draining, the instanceHealthReason value can be one of the following:

Instance.DeregistrationInProgress - The target instance is in the process of being deregistered and the deregistration delay period has not expired.

newInstanceHealthSummary :: InstanceHealthSummary

Create a value of InstanceHealthSummary with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:instanceHealth:InstanceHealthSummary', instanceHealthSummary_instanceHealth - Describes the overall instance health. Valid values are below.

$sel:instanceName:InstanceHealthSummary', instanceHealthSummary_instanceName - The name of the Lightsail instance for which you are requesting health check data.

$sel:instanceHealthReason:InstanceHealthSummary', instanceHealthSummary_instanceHealthReason - More information about the instance health. If the instanceHealth is healthy, then an instanceHealthReason value is not provided.

If instanceHealth is initial, the instanceHealthReason value can be one of the following:

Lb.RegistrationInProgress - The target instance is in the process of being registered with the load balancer.

Lb.InitialHealthChecking - The Lightsail load balancer is still sending the target instance the minimum number of health checks required to determine its health status.

If instanceHealth is unhealthy, the instanceHealthReason value can be one of the following:

Instance.ResponseCodeMismatch - The health checks did not return an expected HTTP code.

Instance.Timeout - The health check requests timed out.

Instance.FailedHealthChecks - The health checks failed because the connection to the target instance timed out, the target instance response was malformed, or the target instance failed the health check for an unknown reason.

Lb.InternalError - The health checks failed due to an internal error.

If instanceHealth is unused, the instanceHealthReason value can be one of the following:

Instance.NotRegistered - The target instance is not registered with the target group.

Instance.NotInUse - The target group is not used by any load balancer, or the target instance is in an Availability Zone that is not enabled for its load balancer.

Instance.IpUnusable - The target IP address is reserved for use by a Lightsail load balancer.

Instance.InvalidState - The target is in the stopped or terminated state.

If instanceHealth is draining, the instanceHealthReason value can be one of the following:

Instance.DeregistrationInProgress - The target instance is in the process of being deregistered and the deregistration delay period has not expired.

instanceHealthSummary_instanceHealth :: Lens' InstanceHealthSummary (Maybe InstanceHealthState)

Describes the overall instance health. Valid values are below.

instanceHealthSummary_instanceName :: Lens' InstanceHealthSummary (Maybe Text)

The name of the Lightsail instance for which you are requesting health check data.

instanceHealthSummary_instanceHealthReason :: Lens' InstanceHealthSummary (Maybe InstanceHealthReason)

More information about the instance health. If the instanceHealth is healthy, then an instanceHealthReason value is not provided.

If instanceHealth is initial, the instanceHealthReason value can be one of the following:

Lb.RegistrationInProgress - The target instance is in the process of being registered with the load balancer.

Lb.InitialHealthChecking - The Lightsail load balancer is still sending the target instance the minimum number of health checks required to determine its health status.

If instanceHealth is unhealthy, the instanceHealthReason value can be one of the following:

Instance.ResponseCodeMismatch - The health checks did not return an expected HTTP code.

Instance.Timeout - The health check requests timed out.

Instance.FailedHealthChecks - The health checks failed because the connection to the target instance timed out, the target instance response was malformed, or the target instance failed the health check for an unknown reason.

Lb.InternalError - The health checks failed due to an internal error.

If instanceHealth is unused, the instanceHealthReason value can be one of the following:

Instance.NotRegistered - The target instance is not registered with the target group.

Instance.NotInUse - The target group is not used by any load balancer, or the target instance is in an Availability Zone that is not enabled for its load balancer.

Instance.IpUnusable - The target IP address is reserved for use by a Lightsail load balancer.

Instance.InvalidState - The target is in the stopped or terminated state.

If instanceHealth is draining, the instanceHealthReason value can be one of the following:

Instance.DeregistrationInProgress - The target instance is in the process of being deregistered and the deregistration delay period has not expired.