Amazonka.Lightsail.Types.InstanceNetworking

Description

Documentation

data InstanceNetworking

Describes monthly data transfer rates and port information for an instance.

See: newInstanceNetworking smart constructor.

Constructors

InstanceNetworking'

Fields

monthlyTransfer :: Maybe MonthlyTransfer

The amount of data in GB allocated for monthly data transfers.

ports :: Maybe [InstancePortInfo]

An array of key-value pairs containing information about the ports on the instance.

newInstanceNetworking :: InstanceNetworking

Create a value of InstanceNetworking with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:monthlyTransfer:InstanceNetworking', instanceNetworking_monthlyTransfer - The amount of data in GB allocated for monthly data transfers.

$sel:ports:InstanceNetworking', instanceNetworking_ports - An array of key-value pairs containing information about the ports on the instance.

instanceNetworking_monthlyTransfer :: Lens' InstanceNetworking (Maybe MonthlyTransfer)

The amount of data in GB allocated for monthly data transfers.

instanceNetworking_ports :: Lens' InstanceNetworking (Maybe [InstancePortInfo])

An array of key-value pairs containing information about the ports on the instance.