Amazonka.Lightsail.Types.InstanceSnapshot

Description

Documentation

data InstanceSnapshot

Describes an instance snapshot.

See: newInstanceSnapshot smart constructor.

Constructors

InstanceSnapshot'

Fields

fromBlueprintId :: Maybe Text

The blueprint ID from which you created the snapshot (e.g., os_debian_8_3). A blueprint is a virtual private server (or instance) image used to create instances quickly.

isFromAutoSnapshot :: Maybe Bool

A Boolean value indicating whether the snapshot was created from an automatic snapshot.

state :: Maybe InstanceSnapshotState

The state the snapshot is in.

resourceType :: Maybe ResourceType

The type of resource (usually InstanceSnapshot).

fromAttachedDisks :: Maybe [Disk]

An array of disk objects containing information about all block storage disks.

arn :: Maybe Text

The Amazon Resource Name (ARN) of the snapshot (e.g., arn:aws:lightsail:us-east-2:123456789101:InstanceSnapshot/d23b5706-3322-4d83-81e5-12345EXAMPLE).

createdAt :: Maybe POSIX

The timestamp when the snapshot was created (e.g., 1479907467.024).

location :: Maybe ResourceLocation

The region name and Availability Zone where you created the snapshot.

progress :: Maybe Text

The progress of the snapshot.

This is populated only for disk snapshots, and is null for instance snapshots.

name :: Maybe Text

The name of the snapshot.

fromBundleId :: Maybe Text

The bundle ID from which you created the snapshot (e.g., micro_1_0).

sizeInGb :: Maybe Int

The size in GB of the SSD.

supportCode :: Maybe Text

The support code. Include this code in your email to support when you have questions about an instance or another resource in Lightsail. This code enables our support team to look up your Lightsail information more easily.

fromInstanceArn :: Maybe Text

The Amazon Resource Name (ARN) of the instance from which the snapshot was created (e.g., arn:aws:lightsail:us-east-2:123456789101:Instance/64b8404c-ccb1-430b-8daf-12345EXAMPLE).

fromInstanceName :: Maybe Text

The instance from which the snapshot was created.

tags :: Maybe [Tag]

The tag keys and optional values for the resource. For more information about tags in Lightsail, see the Amazon Lightsail Developer Guide.

newInstanceSnapshot :: InstanceSnapshot

Create a value of InstanceSnapshot with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:fromBlueprintId:InstanceSnapshot', instanceSnapshot_fromBlueprintId - The blueprint ID from which you created the snapshot (e.g., os_debian_8_3). A blueprint is a virtual private server (or instance) image used to create instances quickly.

$sel:isFromAutoSnapshot:InstanceSnapshot', instanceSnapshot_isFromAutoSnapshot - A Boolean value indicating whether the snapshot was created from an automatic snapshot.

$sel:state:InstanceSnapshot', instanceSnapshot_state - The state the snapshot is in.

$sel:resourceType:InstanceSnapshot', instanceSnapshot_resourceType - The type of resource (usually InstanceSnapshot).

$sel:fromAttachedDisks:InstanceSnapshot', instanceSnapshot_fromAttachedDisks - An array of disk objects containing information about all block storage disks.

$sel:arn:InstanceSnapshot', instanceSnapshot_arn - The Amazon Resource Name (ARN) of the snapshot (e.g., arn:aws:lightsail:us-east-2:123456789101:InstanceSnapshot/d23b5706-3322-4d83-81e5-12345EXAMPLE).

$sel:createdAt:InstanceSnapshot', instanceSnapshot_createdAt - The timestamp when the snapshot was created (e.g., 1479907467.024).

$sel:location:InstanceSnapshot', instanceSnapshot_location - The region name and Availability Zone where you created the snapshot.

$sel:progress:InstanceSnapshot', instanceSnapshot_progress - The progress of the snapshot.

This is populated only for disk snapshots, and is null for instance snapshots.

$sel:name:InstanceSnapshot', instanceSnapshot_name - The name of the snapshot.

$sel:fromBundleId:InstanceSnapshot', instanceSnapshot_fromBundleId - The bundle ID from which you created the snapshot (e.g., micro_1_0).

$sel:sizeInGb:InstanceSnapshot', instanceSnapshot_sizeInGb - The size in GB of the SSD.

$sel:supportCode:InstanceSnapshot', instanceSnapshot_supportCode - The support code. Include this code in your email to support when you have questions about an instance or another resource in Lightsail. This code enables our support team to look up your Lightsail information more easily.

$sel:fromInstanceArn:InstanceSnapshot', instanceSnapshot_fromInstanceArn - The Amazon Resource Name (ARN) of the instance from which the snapshot was created (e.g., arn:aws:lightsail:us-east-2:123456789101:Instance/64b8404c-ccb1-430b-8daf-12345EXAMPLE).

$sel:fromInstanceName:InstanceSnapshot', instanceSnapshot_fromInstanceName - The instance from which the snapshot was created.

$sel:tags:InstanceSnapshot', instanceSnapshot_tags - The tag keys and optional values for the resource. For more information about tags in Lightsail, see the Amazon Lightsail Developer Guide.

instanceSnapshot_fromBlueprintId :: Lens' InstanceSnapshot (Maybe Text)

The blueprint ID from which you created the snapshot (e.g., os_debian_8_3). A blueprint is a virtual private server (or instance) image used to create instances quickly.

instanceSnapshot_isFromAutoSnapshot :: Lens' InstanceSnapshot (Maybe Bool)

A Boolean value indicating whether the snapshot was created from an automatic snapshot.

instanceSnapshot_state :: Lens' InstanceSnapshot (Maybe InstanceSnapshotState)

The state the snapshot is in.

instanceSnapshot_resourceType :: Lens' InstanceSnapshot (Maybe ResourceType)

The type of resource (usually InstanceSnapshot).

instanceSnapshot_fromAttachedDisks :: Lens' InstanceSnapshot (Maybe [Disk])

An array of disk objects containing information about all block storage disks.

instanceSnapshot_arn :: Lens' InstanceSnapshot (Maybe Text)

The Amazon Resource Name (ARN) of the snapshot (e.g., arn:aws:lightsail:us-east-2:123456789101:InstanceSnapshot/d23b5706-3322-4d83-81e5-12345EXAMPLE).

instanceSnapshot_createdAt :: Lens' InstanceSnapshot (Maybe UTCTime)

The timestamp when the snapshot was created (e.g., 1479907467.024).

instanceSnapshot_location :: Lens' InstanceSnapshot (Maybe ResourceLocation)

The region name and Availability Zone where you created the snapshot.

instanceSnapshot_progress :: Lens' InstanceSnapshot (Maybe Text)

The progress of the snapshot.

This is populated only for disk snapshots, and is null for instance snapshots.

instanceSnapshot_name :: Lens' InstanceSnapshot (Maybe Text)

The name of the snapshot.

instanceSnapshot_fromBundleId :: Lens' InstanceSnapshot (Maybe Text)

The bundle ID from which you created the snapshot (e.g., micro_1_0).

instanceSnapshot_sizeInGb :: Lens' InstanceSnapshot (Maybe Int)

The size in GB of the SSD.

instanceSnapshot_supportCode :: Lens' InstanceSnapshot (Maybe Text)

The support code. Include this code in your email to support when you have questions about an instance or another resource in Lightsail. This code enables our support team to look up your Lightsail information more easily.

instanceSnapshot_fromInstanceArn :: Lens' InstanceSnapshot (Maybe Text)

The Amazon Resource Name (ARN) of the instance from which the snapshot was created (e.g., arn:aws:lightsail:us-east-2:123456789101:Instance/64b8404c-ccb1-430b-8daf-12345EXAMPLE).

instanceSnapshot_fromInstanceName :: Lens' InstanceSnapshot (Maybe Text)

The instance from which the snapshot was created.

instanceSnapshot_tags :: Lens' InstanceSnapshot (Maybe [Tag])

The tag keys and optional values for the resource. For more information about tags in Lightsail, see the Amazon Lightsail Developer Guide.