Amazonka.Lightsail.Types.RelationalDatabase

Description

Documentation

data RelationalDatabase

Describes a database.

See: newRelationalDatabase smart constructor.

Constructors

RelationalDatabase'

Fields

engineVersion :: Maybe Text

The database engine version (for example, 5.7.23).

relationalDatabaseBundleId :: Maybe Text

The bundle ID for the database. A bundle describes the performance specifications for your database.

masterEndpoint :: Maybe RelationalDatabaseEndpoint

The master endpoint for the database.

state :: Maybe Text

Describes the current state of the database.

resourceType :: Maybe ResourceType

The Lightsail resource type for the database (for example, RelationalDatabase).

publiclyAccessible :: Maybe Bool

A Boolean value indicating whether the database is publicly accessible.

masterUsername :: Maybe Text

The master user name of the database.

arn :: Maybe Text

The Amazon Resource Name (ARN) of the database.

createdAt :: Maybe POSIX

The timestamp when the database was created. Formatted in Unix time.

location :: Maybe ResourceLocation

The Region name and Availability Zone where the database is located.

engine :: Maybe Text

The database software (for example, MySQL).

latestRestorableTime :: Maybe POSIX

The latest point in time to which the database can be restored. Formatted in Unix time.

preferredMaintenanceWindow :: Maybe Text

The weekly time range during which system maintenance can occur on the database.

In the format ddd:hh24:mi-ddd:hh24:mi. For example, Tue:17:00-Tue:17:30.

relationalDatabaseBlueprintId :: Maybe Text

The blueprint ID for the database. A blueprint describes the major engine version of a database.

caCertificateIdentifier :: Maybe Text

The certificate associated with the database.

name :: Maybe Text

The unique name of the database resource in Lightsail.

backupRetentionEnabled :: Maybe Bool

A Boolean value indicating whether automated backup retention is enabled for the database.

preferredBackupWindow :: Maybe Text

The daily time range during which automated backups are created for the database (for example, 16:00-16:30).

pendingMaintenanceActions :: Maybe [PendingMaintenanceAction]

Describes the pending maintenance actions for the database.

supportCode :: Maybe Text

The support code for the database. Include this code in your email to support when you have questions about a database in Lightsail. This code enables our support team to look up your Lightsail information more easily.

secondaryAvailabilityZone :: Maybe Text

Describes the secondary Availability Zone of a high availability database.

The secondary database is used for failover support of a high availability database.

pendingModifiedValues :: Maybe PendingModifiedRelationalDatabaseValues

Describes pending database value modifications.

masterDatabaseName :: Maybe Text

The name of the master database created when the Lightsail database resource is created.

hardware :: Maybe RelationalDatabaseHardware

Describes the hardware of the database.

parameterApplyStatus :: Maybe Text

The status of parameter updates for the database.

tags :: Maybe [Tag]

The tag keys and optional values for the resource. For more information about tags in Lightsail, see the Amazon Lightsail Developer Guide.

newRelationalDatabase :: RelationalDatabase

Create a value of RelationalDatabase with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:engineVersion:RelationalDatabase', relationalDatabase_engineVersion - The database engine version (for example, 5.7.23).

$sel:relationalDatabaseBundleId:RelationalDatabase', relationalDatabase_relationalDatabaseBundleId - The bundle ID for the database. A bundle describes the performance specifications for your database.

$sel:masterEndpoint:RelationalDatabase', relationalDatabase_masterEndpoint - The master endpoint for the database.

$sel:state:RelationalDatabase', relationalDatabase_state - Describes the current state of the database.

$sel:resourceType:RelationalDatabase', relationalDatabase_resourceType - The Lightsail resource type for the database (for example, RelationalDatabase).

$sel:publiclyAccessible:RelationalDatabase', relationalDatabase_publiclyAccessible - A Boolean value indicating whether the database is publicly accessible.

$sel:masterUsername:RelationalDatabase', relationalDatabase_masterUsername - The master user name of the database.

$sel:arn:RelationalDatabase', relationalDatabase_arn - The Amazon Resource Name (ARN) of the database.

$sel:createdAt:RelationalDatabase', relationalDatabase_createdAt - The timestamp when the database was created. Formatted in Unix time.

$sel:location:RelationalDatabase', relationalDatabase_location - The Region name and Availability Zone where the database is located.

$sel:engine:RelationalDatabase', relationalDatabase_engine - The database software (for example, MySQL).

$sel:latestRestorableTime:RelationalDatabase', relationalDatabase_latestRestorableTime - The latest point in time to which the database can be restored. Formatted in Unix time.

$sel:preferredMaintenanceWindow:RelationalDatabase', relationalDatabase_preferredMaintenanceWindow - The weekly time range during which system maintenance can occur on the database.

In the format ddd:hh24:mi-ddd:hh24:mi. For example, Tue:17:00-Tue:17:30.

$sel:relationalDatabaseBlueprintId:RelationalDatabase', relationalDatabase_relationalDatabaseBlueprintId - The blueprint ID for the database. A blueprint describes the major engine version of a database.

$sel:caCertificateIdentifier:RelationalDatabase', relationalDatabase_caCertificateIdentifier - The certificate associated with the database.

$sel:name:RelationalDatabase', relationalDatabase_name - The unique name of the database resource in Lightsail.

$sel:backupRetentionEnabled:RelationalDatabase', relationalDatabase_backupRetentionEnabled - A Boolean value indicating whether automated backup retention is enabled for the database.

$sel:preferredBackupWindow:RelationalDatabase', relationalDatabase_preferredBackupWindow - The daily time range during which automated backups are created for the database (for example, 16:00-16:30).

$sel:pendingMaintenanceActions:RelationalDatabase', relationalDatabase_pendingMaintenanceActions - Describes the pending maintenance actions for the database.

$sel:supportCode:RelationalDatabase', relationalDatabase_supportCode - The support code for the database. Include this code in your email to support when you have questions about a database in Lightsail. This code enables our support team to look up your Lightsail information more easily.

$sel:secondaryAvailabilityZone:RelationalDatabase', relationalDatabase_secondaryAvailabilityZone - Describes the secondary Availability Zone of a high availability database.

The secondary database is used for failover support of a high availability database.

$sel:pendingModifiedValues:RelationalDatabase', relationalDatabase_pendingModifiedValues - Describes pending database value modifications.

$sel:masterDatabaseName:RelationalDatabase', relationalDatabase_masterDatabaseName - The name of the master database created when the Lightsail database resource is created.

$sel:hardware:RelationalDatabase', relationalDatabase_hardware - Describes the hardware of the database.

$sel:parameterApplyStatus:RelationalDatabase', relationalDatabase_parameterApplyStatus - The status of parameter updates for the database.

$sel:tags:RelationalDatabase', relationalDatabase_tags - The tag keys and optional values for the resource. For more information about tags in Lightsail, see the Amazon Lightsail Developer Guide.

relationalDatabase_engineVersion :: Lens' RelationalDatabase (Maybe Text)

The database engine version (for example, 5.7.23).

relationalDatabase_relationalDatabaseBundleId :: Lens' RelationalDatabase (Maybe Text)

The bundle ID for the database. A bundle describes the performance specifications for your database.

relationalDatabase_masterEndpoint :: Lens' RelationalDatabase (Maybe RelationalDatabaseEndpoint)

The master endpoint for the database.

relationalDatabase_state :: Lens' RelationalDatabase (Maybe Text)

Describes the current state of the database.

relationalDatabase_resourceType :: Lens' RelationalDatabase (Maybe ResourceType)

The Lightsail resource type for the database (for example, RelationalDatabase).

relationalDatabase_publiclyAccessible :: Lens' RelationalDatabase (Maybe Bool)

A Boolean value indicating whether the database is publicly accessible.

relationalDatabase_masterUsername :: Lens' RelationalDatabase (Maybe Text)

The master user name of the database.

relationalDatabase_arn :: Lens' RelationalDatabase (Maybe Text)

The Amazon Resource Name (ARN) of the database.

relationalDatabase_createdAt :: Lens' RelationalDatabase (Maybe UTCTime)

The timestamp when the database was created. Formatted in Unix time.

relationalDatabase_location :: Lens' RelationalDatabase (Maybe ResourceLocation)

The Region name and Availability Zone where the database is located.

relationalDatabase_engine :: Lens' RelationalDatabase (Maybe Text)

The database software (for example, MySQL).

relationalDatabase_latestRestorableTime :: Lens' RelationalDatabase (Maybe UTCTime)

The latest point in time to which the database can be restored. Formatted in Unix time.

relationalDatabase_preferredMaintenanceWindow :: Lens' RelationalDatabase (Maybe Text)

The weekly time range during which system maintenance can occur on the database.

In the format ddd:hh24:mi-ddd:hh24:mi. For example, Tue:17:00-Tue:17:30.

relationalDatabase_relationalDatabaseBlueprintId :: Lens' RelationalDatabase (Maybe Text)

The blueprint ID for the database. A blueprint describes the major engine version of a database.

relationalDatabase_caCertificateIdentifier :: Lens' RelationalDatabase (Maybe Text)

The certificate associated with the database.

relationalDatabase_name :: Lens' RelationalDatabase (Maybe Text)

The unique name of the database resource in Lightsail.

relationalDatabase_backupRetentionEnabled :: Lens' RelationalDatabase (Maybe Bool)

A Boolean value indicating whether automated backup retention is enabled for the database.

relationalDatabase_preferredBackupWindow :: Lens' RelationalDatabase (Maybe Text)

The daily time range during which automated backups are created for the database (for example, 16:00-16:30).

relationalDatabase_pendingMaintenanceActions :: Lens' RelationalDatabase (Maybe [PendingMaintenanceAction])

Describes the pending maintenance actions for the database.

relationalDatabase_supportCode :: Lens' RelationalDatabase (Maybe Text)

The support code for the database. Include this code in your email to support when you have questions about a database in Lightsail. This code enables our support team to look up your Lightsail information more easily.

relationalDatabase_secondaryAvailabilityZone :: Lens' RelationalDatabase (Maybe Text)

Describes the secondary Availability Zone of a high availability database.

The secondary database is used for failover support of a high availability database.

relationalDatabase_pendingModifiedValues :: Lens' RelationalDatabase (Maybe PendingModifiedRelationalDatabaseValues)

Describes pending database value modifications.

relationalDatabase_masterDatabaseName :: Lens' RelationalDatabase (Maybe Text)

The name of the master database created when the Lightsail database resource is created.

relationalDatabase_hardware :: Lens' RelationalDatabase (Maybe RelationalDatabaseHardware)

Describes the hardware of the database.

relationalDatabase_parameterApplyStatus :: Lens' RelationalDatabase (Maybe Text)

The status of parameter updates for the database.

relationalDatabase_tags :: Lens' RelationalDatabase (Maybe [Tag])

The tag keys and optional values for the resource. For more information about tags in Lightsail, see the Amazon Lightsail Developer Guide.