Amazonka.Lightsail.UpdateBucketBundle

Description

Updates the bundle, or storage plan, of an existing Amazon Lightsail bucket.

A bucket bundle specifies the monthly cost, storage space, and data transfer quota for a bucket. You can update a bucket's bundle only one time within a monthly AWS billing cycle. To determine if you can update a bucket's bundle, use the GetBuckets action. The ableToUpdateBundle parameter in the response will indicate whether you can currently update a bucket's bundle.

Update a bucket's bundle if it's consistently going over its storage space or data transfer quota, or if a bucket's usage is consistently in the lower range of its storage space or data transfer quota. Due to the unpredictable usage fluctuations that a bucket might experience, we strongly recommend that you update a bucket's bundle only as a long-term strategy, instead of as a short-term, monthly cost-cutting measure. Choose a bucket bundle that will provide the bucket with ample storage space and data transfer for a long time to come.

Creating a Request

data UpdateBucketBundle

See: newUpdateBucketBundle smart constructor.

Constructors

UpdateBucketBundle'

Fields

bucketName :: Text

The name of the bucket for which to update the bundle.

bundleId :: Text

The ID of the new bundle to apply to the bucket.

Use the GetBucketBundles action to get a list of bundle IDs that you can specify.

newUpdateBucketBundle

Arguments

:: Text
$sel:bucketName:UpdateBucketBundle'

-> Text
$sel:bundleId:UpdateBucketBundle'

-> UpdateBucketBundle

Create a value of UpdateBucketBundle with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:bucketName:UpdateBucketBundle', updateBucketBundle_bucketName - The name of the bucket for which to update the bundle.

$sel:bundleId:UpdateBucketBundle', updateBucketBundle_bundleId - The ID of the new bundle to apply to the bucket.

Use the GetBucketBundles action to get a list of bundle IDs that you can specify.

Request Lenses

updateBucketBundle_bucketName :: Lens' UpdateBucketBundle Text

The name of the bucket for which to update the bundle.

updateBucketBundle_bundleId :: Lens' UpdateBucketBundle Text

The ID of the new bundle to apply to the bucket.

Use the GetBucketBundles action to get a list of bundle IDs that you can specify.

Destructuring the Response

data UpdateBucketBundleResponse

See: newUpdateBucketBundleResponse smart constructor.

Constructors

UpdateBucketBundleResponse'

Fields

operations :: Maybe [Operation]

An array of objects that describe the result of the action, such as the status of the request, the timestamp of the request, and the resources affected by the request.

httpStatus :: Int

The response's http status code.

newUpdateBucketBundleResponse

Arguments

:: Int
$sel:httpStatus:UpdateBucketBundleResponse'

-> UpdateBucketBundleResponse

Create a value of UpdateBucketBundleResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:operations:UpdateBucketBundleResponse', updateBucketBundleResponse_operations - An array of objects that describe the result of the action, such as the status of the request, the timestamp of the request, and the resources affected by the request.

$sel:httpStatus:UpdateBucketBundleResponse', updateBucketBundleResponse_httpStatus - The response's http status code.

Response Lenses

updateBucketBundleResponse_operations :: Lens' UpdateBucketBundleResponse (Maybe [Operation])

An array of objects that describe the result of the action, such as the status of the request, the timestamp of the request, and the resources affected by the request.

updateBucketBundleResponse_httpStatus :: Lens' UpdateBucketBundleResponse Int

The response's http status code.