Amazonka.Location.BatchGetDevicePosition

Description

Lists the latest device positions for requested devices.

Creating a Request

data BatchGetDevicePosition

See: newBatchGetDevicePosition smart constructor.

Constructors

BatchGetDevicePosition'

Fields

deviceIds :: NonEmpty Text

Devices whose position you want to retrieve.

For example, for two devices: device-ids=DeviceId1&device-ids=DeviceId2

trackerName :: Text

The tracker resource retrieving the device position.

newBatchGetDevicePosition

Arguments

:: NonEmpty Text
$sel:deviceIds:BatchGetDevicePosition'

-> Text
$sel:trackerName:BatchGetDevicePosition'

-> BatchGetDevicePosition

Create a value of BatchGetDevicePosition with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:deviceIds:BatchGetDevicePosition', batchGetDevicePosition_deviceIds - Devices whose position you want to retrieve.

For example, for two devices: device-ids=DeviceId1&device-ids=DeviceId2

$sel:trackerName:BatchGetDevicePosition', batchGetDevicePosition_trackerName - The tracker resource retrieving the device position.

Request Lenses

batchGetDevicePosition_deviceIds :: Lens' BatchGetDevicePosition (NonEmpty Text)

Devices whose position you want to retrieve.

For example, for two devices: device-ids=DeviceId1&device-ids=DeviceId2

batchGetDevicePosition_trackerName :: Lens' BatchGetDevicePosition Text

The tracker resource retrieving the device position.

Destructuring the Response

data BatchGetDevicePositionResponse

See: newBatchGetDevicePositionResponse smart constructor.

Constructors

BatchGetDevicePositionResponse'

Fields

httpStatus :: Int

The response's http status code.

devicePositions :: [DevicePosition]

Contains device position details such as the device ID, position, and timestamps for when the position was received and sampled.

errors :: [BatchGetDevicePositionError]

Contains error details for each device that failed to send its position to the tracker resource.

newBatchGetDevicePositionResponse

Arguments

:: Int
$sel:httpStatus:BatchGetDevicePositionResponse'

-> BatchGetDevicePositionResponse

Create a value of BatchGetDevicePositionResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:httpStatus:BatchGetDevicePositionResponse', batchGetDevicePositionResponse_httpStatus - The response's http status code.

$sel:devicePositions:BatchGetDevicePositionResponse', batchGetDevicePositionResponse_devicePositions - Contains device position details such as the device ID, position, and timestamps for when the position was received and sampled.

$sel:errors:BatchGetDevicePositionResponse', batchGetDevicePositionResponse_errors - Contains error details for each device that failed to send its position to the tracker resource.

Response Lenses

batchGetDevicePositionResponse_httpStatus :: Lens' BatchGetDevicePositionResponse Int

The response's http status code.

batchGetDevicePositionResponse_devicePositions :: Lens' BatchGetDevicePositionResponse [DevicePosition]

Contains device position details such as the device ID, position, and timestamps for when the position was received and sampled.

batchGetDevicePositionResponse_errors :: Lens' BatchGetDevicePositionResponse [BatchGetDevicePositionError]

Contains error details for each device that failed to send its position to the tracker resource.