Amazonka.Location.CreateGeofenceCollection

Description

Creates a geofence collection, which manages and stores geofences.

Creating a Request

data CreateGeofenceCollection

See: newCreateGeofenceCollection smart constructor.

Constructors

CreateGeofenceCollection'

Fields

pricingPlanDataSource :: Maybe Text

Specifies the data provider for the geofence collection.

Required value for the following pricing plans: MobileAssetTracking | MobileAssetManagement

For more information about Data Providers, and Pricing plans, see the Amazon Location Service product page.

Amazon Location Service only uses PricingPlanDataSource to calculate billing for your geofence collection. Your data won't be shared with the data provider, and will remain in your AWS account or Region unless you move it.

Valid Values: Esri | Here

kmsKeyId :: Maybe Text

A key identifier for an AWS KMS customer managed key. Enter a key ID, key ARN, alias name, or alias ARN.

description :: Maybe Text

An optional description for the geofence collection.

tags :: Maybe (HashMap Text Text)

Applies one or more tags to the geofence collection. A tag is a key-value pair helps manage, identify, search, and filter your resources by labelling them.

Format: "key" : "value"

Restrictions:

Maximum 50 tags per resource

Each resource tag must be unique with a maximum of one value.

Maximum key length: 128 Unicode characters in UTF-8

Maximum value length: 256 Unicode characters in UTF-8

Can use alphanumeric characters (A–Z, a–z, 0–9), and the following characters: + - = . _ : / @.

collectionName :: Text

A custom name for the geofence collection.

Requirements:

Contain only alphanumeric characters (A–Z, a–z, 0–9), hyphens (-), periods (.), and underscores (_).

Must be a unique geofence collection name.

No spaces allowed. For example, ExampleGeofenceCollection.

pricingPlan :: PricingPlan

Specifies the pricing plan for the geofence collection.

For additional details and restrictions on each pricing plan option, see the Amazon Location Service pricing page.

newCreateGeofenceCollection

Arguments

:: Text
$sel:collectionName:CreateGeofenceCollection'

-> PricingPlan
$sel:pricingPlan:CreateGeofenceCollection'

-> CreateGeofenceCollection

Create a value of CreateGeofenceCollection with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:pricingPlanDataSource:CreateGeofenceCollection', createGeofenceCollection_pricingPlanDataSource - Specifies the data provider for the geofence collection.

Required value for the following pricing plans: MobileAssetTracking | MobileAssetManagement

For more information about Data Providers, and Pricing plans, see the Amazon Location Service product page.

Amazon Location Service only uses PricingPlanDataSource to calculate billing for your geofence collection. Your data won't be shared with the data provider, and will remain in your AWS account or Region unless you move it.

Valid Values: Esri | Here

$sel:kmsKeyId:CreateGeofenceCollection', createGeofenceCollection_kmsKeyId - A key identifier for an AWS KMS customer managed key. Enter a key ID, key ARN, alias name, or alias ARN.

$sel:description:CreateGeofenceCollection', createGeofenceCollection_description - An optional description for the geofence collection.

$sel:tags:CreateGeofenceCollection', createGeofenceCollection_tags - Applies one or more tags to the geofence collection. A tag is a key-value pair helps manage, identify, search, and filter your resources by labelling them.

Format: "key" : "value"

Restrictions:

Maximum 50 tags per resource

Each resource tag must be unique with a maximum of one value.

Maximum key length: 128 Unicode characters in UTF-8

Maximum value length: 256 Unicode characters in UTF-8

Can use alphanumeric characters (A–Z, a–z, 0–9), and the following characters: + - = . _ : / @.

$sel:collectionName:CreateGeofenceCollection', createGeofenceCollection_collectionName - A custom name for the geofence collection.

Requirements:

Contain only alphanumeric characters (A–Z, a–z, 0–9), hyphens (-), periods (.), and underscores (_).

Must be a unique geofence collection name.

No spaces allowed. For example, ExampleGeofenceCollection.

$sel:pricingPlan:CreateGeofenceCollection', createGeofenceCollection_pricingPlan - Specifies the pricing plan for the geofence collection.

For additional details and restrictions on each pricing plan option, see the Amazon Location Service pricing page.

Request Lenses

createGeofenceCollection_pricingPlanDataSource :: Lens' CreateGeofenceCollection (Maybe Text)

Specifies the data provider for the geofence collection.

Required value for the following pricing plans: MobileAssetTracking | MobileAssetManagement

For more information about Data Providers, and Pricing plans, see the Amazon Location Service product page.

Amazon Location Service only uses PricingPlanDataSource to calculate billing for your geofence collection. Your data won't be shared with the data provider, and will remain in your AWS account or Region unless you move it.

Valid Values: Esri | Here

createGeofenceCollection_kmsKeyId :: Lens' CreateGeofenceCollection (Maybe Text)

A key identifier for an AWS KMS customer managed key. Enter a key ID, key ARN, alias name, or alias ARN.

createGeofenceCollection_description :: Lens' CreateGeofenceCollection (Maybe Text)

An optional description for the geofence collection.

createGeofenceCollection_tags :: Lens' CreateGeofenceCollection (Maybe (HashMap Text Text))

Applies one or more tags to the geofence collection. A tag is a key-value pair helps manage, identify, search, and filter your resources by labelling them.

Format: "key" : "value"

Restrictions:

Maximum 50 tags per resource

Each resource tag must be unique with a maximum of one value.

Maximum key length: 128 Unicode characters in UTF-8

Maximum value length: 256 Unicode characters in UTF-8

Can use alphanumeric characters (A–Z, a–z, 0–9), and the following characters: + - = . _ : / @.

createGeofenceCollection_collectionName :: Lens' CreateGeofenceCollection Text

A custom name for the geofence collection.

Requirements:

Contain only alphanumeric characters (A–Z, a–z, 0–9), hyphens (-), periods (.), and underscores (_).

Must be a unique geofence collection name.

No spaces allowed. For example, ExampleGeofenceCollection.

createGeofenceCollection_pricingPlan :: Lens' CreateGeofenceCollection PricingPlan

Specifies the pricing plan for the geofence collection.

For additional details and restrictions on each pricing plan option, see the Amazon Location Service pricing page.

Destructuring the Response

data CreateGeofenceCollectionResponse

See: newCreateGeofenceCollectionResponse smart constructor.

Constructors

CreateGeofenceCollectionResponse'

Fields

httpStatus :: Int

The response's http status code.

collectionArn :: Text

The Amazon Resource Name (ARN) for the geofence collection resource. Used when you need to specify a resource across all AWS.

Format example: arn:aws:geo:region:account-id:geofence-collection/ExampleGeofenceCollection

collectionName :: Text

The name for the geofence collection.

createTime :: POSIX

The timestamp for when the geofence collection was created in ISO 8601 format: YYYY-MM-DDThh:mm:ss.sssZ

newCreateGeofenceCollectionResponse

Arguments

:: Int
$sel:httpStatus:CreateGeofenceCollectionResponse'

-> Text
$sel:collectionArn:CreateGeofenceCollectionResponse'

-> Text
$sel:collectionName:CreateGeofenceCollection'

-> UTCTime
$sel:createTime:CreateGeofenceCollectionResponse'

-> CreateGeofenceCollectionResponse

Create a value of CreateGeofenceCollectionResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:httpStatus:CreateGeofenceCollectionResponse', createGeofenceCollectionResponse_httpStatus - The response's http status code.

$sel:collectionArn:CreateGeofenceCollectionResponse', createGeofenceCollectionResponse_collectionArn - The Amazon Resource Name (ARN) for the geofence collection resource. Used when you need to specify a resource across all AWS.

Format example: arn:aws:geo:region:account-id:geofence-collection/ExampleGeofenceCollection

$sel:collectionName:CreateGeofenceCollection', createGeofenceCollectionResponse_collectionName - The name for the geofence collection.

$sel:createTime:CreateGeofenceCollectionResponse', createGeofenceCollectionResponse_createTime - The timestamp for when the geofence collection was created in ISO 8601 format: YYYY-MM-DDThh:mm:ss.sssZ

Response Lenses

createGeofenceCollectionResponse_httpStatus :: Lens' CreateGeofenceCollectionResponse Int

The response's http status code.

createGeofenceCollectionResponse_collectionArn :: Lens' CreateGeofenceCollectionResponse Text

The Amazon Resource Name (ARN) for the geofence collection resource. Used when you need to specify a resource across all AWS.

Format example: arn:aws:geo:region:account-id:geofence-collection/ExampleGeofenceCollection

createGeofenceCollectionResponse_collectionName :: Lens' CreateGeofenceCollectionResponse Text

The name for the geofence collection.

createGeofenceCollectionResponse_createTime :: Lens' CreateGeofenceCollectionResponse UTCTime

The timestamp for when the geofence collection was created in ISO 8601 format: YYYY-MM-DDThh:mm:ss.sssZ