Amazonka.Location.CreatePlaceIndex

Description

Creates a place index resource in your AWS account, which supports functions with geospatial data sourced from your chosen data provider.

Creating a Request

data CreatePlaceIndex

See: newCreatePlaceIndex smart constructor.

Constructors

CreatePlaceIndex'

Fields

dataSourceConfiguration :: Maybe DataSourceConfiguration

Specifies the data storage option requesting Places.

description :: Maybe Text

The optional description for the place index resource.

tags :: Maybe (HashMap Text Text)

Applies one or more tags to the place index resource. A tag is a key-value pair helps manage, identify, search, and filter your resources by labelling them.

Format: "key" : "value"

Restrictions:

Maximum 50 tags per resource

Each resource tag must be unique with a maximum of one value.

Maximum key length: 128 Unicode characters in UTF-8

Maximum value length: 256 Unicode characters in UTF-8

Can use alphanumeric characters (A–Z, a–z, 0–9), and the following characters: + - = . _ : / @.

dataSource :: Text

Specifies the data provider of geospatial data.

This field is case-sensitive. Enter the valid values as shown. For example, entering HERE returns an error.

Valid values include:

Esri – For additional information about Esri's coverage in your region of interest, see Esri details on geocoding coverage.

Here – For additional information about HERE Technologies' coverage in your region of interest, see HERE details on goecoding coverage.

Place index resources using HERE Technologies as a data provider can't store results for locations in Japan. For more information, see the AWS Service Terms for Amazon Location Service.

For additional information , see Data providers on the Amazon Location Service Developer Guide.

indexName :: Text

The name of the place index resource.

Requirements:

Contain only alphanumeric characters (A–Z, a–z, 0–9), hyphens (-), periods (.), and underscores (_).

Must be a unique place index resource name.

No spaces allowed. For example, ExamplePlaceIndex.

pricingPlan :: PricingPlan

Specifies the pricing plan for your place index resource.

For additional details and restrictions on each pricing plan option, see Amazon Location Service pricing.

newCreatePlaceIndex

Arguments

:: Text
$sel:dataSource:CreatePlaceIndex'

-> Text
$sel:indexName:CreatePlaceIndex'

-> PricingPlan
$sel:pricingPlan:CreatePlaceIndex'

-> CreatePlaceIndex

Create a value of CreatePlaceIndex with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:dataSourceConfiguration:CreatePlaceIndex', createPlaceIndex_dataSourceConfiguration - Specifies the data storage option requesting Places.

$sel:description:CreatePlaceIndex', createPlaceIndex_description - The optional description for the place index resource.

$sel:tags:CreatePlaceIndex', createPlaceIndex_tags - Applies one or more tags to the place index resource. A tag is a key-value pair helps manage, identify, search, and filter your resources by labelling them.

Format: "key" : "value"

Restrictions:

Maximum 50 tags per resource

Each resource tag must be unique with a maximum of one value.

Maximum key length: 128 Unicode characters in UTF-8

Maximum value length: 256 Unicode characters in UTF-8

Can use alphanumeric characters (A–Z, a–z, 0–9), and the following characters: + - = . _ : / @.

$sel:dataSource:CreatePlaceIndex', createPlaceIndex_dataSource - Specifies the data provider of geospatial data.

This field is case-sensitive. Enter the valid values as shown. For example, entering HERE returns an error.

Valid values include:

Esri – For additional information about Esri's coverage in your region of interest, see Esri details on geocoding coverage.

Here – For additional information about HERE Technologies' coverage in your region of interest, see HERE details on goecoding coverage.

Place index resources using HERE Technologies as a data provider can't store results for locations in Japan. For more information, see the AWS Service Terms for Amazon Location Service.

For additional information , see Data providers on the Amazon Location Service Developer Guide.

$sel:indexName:CreatePlaceIndex', createPlaceIndex_indexName - The name of the place index resource.

Requirements:

Contain only alphanumeric characters (A–Z, a–z, 0–9), hyphens (-), periods (.), and underscores (_).

Must be a unique place index resource name.

No spaces allowed. For example, ExamplePlaceIndex.

$sel:pricingPlan:CreatePlaceIndex', createPlaceIndex_pricingPlan - Specifies the pricing plan for your place index resource.

For additional details and restrictions on each pricing plan option, see Amazon Location Service pricing.

Request Lenses

createPlaceIndex_dataSourceConfiguration :: Lens' CreatePlaceIndex (Maybe DataSourceConfiguration)

Specifies the data storage option requesting Places.

createPlaceIndex_description :: Lens' CreatePlaceIndex (Maybe Text)

The optional description for the place index resource.

createPlaceIndex_tags :: Lens' CreatePlaceIndex (Maybe (HashMap Text Text))

Applies one or more tags to the place index resource. A tag is a key-value pair helps manage, identify, search, and filter your resources by labelling them.

Format: "key" : "value"

Restrictions:

Maximum 50 tags per resource

Each resource tag must be unique with a maximum of one value.

Maximum key length: 128 Unicode characters in UTF-8

Maximum value length: 256 Unicode characters in UTF-8

Can use alphanumeric characters (A–Z, a–z, 0–9), and the following characters: + - = . _ : / @.

createPlaceIndex_dataSource :: Lens' CreatePlaceIndex Text

Specifies the data provider of geospatial data.

This field is case-sensitive. Enter the valid values as shown. For example, entering HERE returns an error.

Valid values include:

Esri – For additional information about Esri's coverage in your region of interest, see Esri details on geocoding coverage.

Here – For additional information about HERE Technologies' coverage in your region of interest, see HERE details on goecoding coverage.

Place index resources using HERE Technologies as a data provider can't store results for locations in Japan. For more information, see the AWS Service Terms for Amazon Location Service.

For additional information , see Data providers on the Amazon Location Service Developer Guide.

createPlaceIndex_indexName :: Lens' CreatePlaceIndex Text

The name of the place index resource.

Requirements:

Contain only alphanumeric characters (A–Z, a–z, 0–9), hyphens (-), periods (.), and underscores (_).

Must be a unique place index resource name.

No spaces allowed. For example, ExamplePlaceIndex.

createPlaceIndex_pricingPlan :: Lens' CreatePlaceIndex PricingPlan

Specifies the pricing plan for your place index resource.

For additional details and restrictions on each pricing plan option, see Amazon Location Service pricing.

Destructuring the Response

data CreatePlaceIndexResponse

See: newCreatePlaceIndexResponse smart constructor.

Constructors

CreatePlaceIndexResponse'

Fields

httpStatus :: Int

The response's http status code.

createTime :: POSIX

The timestamp for when the place index resource was created in ISO 8601 format: YYYY-MM-DDThh:mm:ss.sssZ.

indexArn :: Text

The Amazon Resource Name (ARN) for the place index resource. Used to specify a resource across AWS.

Format example: arn:aws:geo:region:account-id:place-index/ExamplePlaceIndex

indexName :: Text

The name for the place index resource.

newCreatePlaceIndexResponse

Arguments

:: Int
$sel:httpStatus:CreatePlaceIndexResponse'

-> UTCTime
$sel:createTime:CreatePlaceIndexResponse'

-> Text
$sel:indexArn:CreatePlaceIndexResponse'

-> Text
$sel:indexName:CreatePlaceIndex'

-> CreatePlaceIndexResponse

Create a value of CreatePlaceIndexResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:httpStatus:CreatePlaceIndexResponse', createPlaceIndexResponse_httpStatus - The response's http status code.

$sel:createTime:CreatePlaceIndexResponse', createPlaceIndexResponse_createTime - The timestamp for when the place index resource was created in ISO 8601 format: YYYY-MM-DDThh:mm:ss.sssZ.

$sel:indexArn:CreatePlaceIndexResponse', createPlaceIndexResponse_indexArn - The Amazon Resource Name (ARN) for the place index resource. Used to specify a resource across AWS.

Format example: arn:aws:geo:region:account-id:place-index/ExamplePlaceIndex

$sel:indexName:CreatePlaceIndex', createPlaceIndexResponse_indexName - The name for the place index resource.

Response Lenses

createPlaceIndexResponse_httpStatus :: Lens' CreatePlaceIndexResponse Int

The response's http status code.

createPlaceIndexResponse_createTime :: Lens' CreatePlaceIndexResponse UTCTime

The timestamp for when the place index resource was created in ISO 8601 format: YYYY-MM-DDThh:mm:ss.sssZ.

createPlaceIndexResponse_indexArn :: Lens' CreatePlaceIndexResponse Text

The Amazon Resource Name (ARN) for the place index resource. Used to specify a resource across AWS.

Format example: arn:aws:geo:region:account-id:place-index/ExamplePlaceIndex

createPlaceIndexResponse_indexName :: Lens' CreatePlaceIndexResponse Text

The name for the place index resource.