Amazonka.Location.CreateRouteCalculator

Description

Creates a route calculator resource in your AWS account.

You can send requests to a route calculator resource to estimate travel time, distance, and get directions. A route calculator sources traffic and road network data from your chosen data provider.

Creating a Request

data CreateRouteCalculator

See: newCreateRouteCalculator smart constructor.

Constructors

CreateRouteCalculator'

Fields

description :: Maybe Text

The optional description for the route calculator resource.

tags :: Maybe (HashMap Text Text)

Applies one or more tags to the route calculator resource. A tag is a key-value pair helps manage, identify, search, and filter your resources by labelling them.

For example: { "tag1" : "value1", "tag2" : "value2"}

Format: "key" : "value"

Restrictions:

Maximum 50 tags per resource

Each resource tag must be unique with a maximum of one value.

Maximum key length: 128 Unicode characters in UTF-8

Maximum value length: 256 Unicode characters in UTF-8

Can use alphanumeric characters (A–Z, a–z, 0–9), and the following characters: + - = . _ : / @.

calculatorName :: Text

The name of the route calculator resource.

Requirements:

Can use alphanumeric characters (A–Z, a–z, 0–9) , hyphens (-), periods (.), and underscores (_).

Must be a unique Route calculator resource name.

No spaces allowed. For example, ExampleRouteCalculator.

dataSource :: Text

Specifies the data provider of traffic and road network data.

This field is case-sensitive. Enter the valid values as shown. For example, entering HERE returns an error. Route calculators that use Esri as a data source only calculate routes that are shorter than 400 km.

Valid values include:

Esri – For additional information about Esri's coverage in your region of interest, see Esri details on street networks and traffic coverage.

Here – For additional information about HERE Technologies' coverage in your region of interest, see HERE car routing coverage and HERE truck routing coverage.

For additional information , see Data providers on the Amazon Location Service Developer Guide.

pricingPlan :: PricingPlan

Specifies the pricing plan for your route calculator resource.

For additional details and restrictions on each pricing plan option, see Amazon Location Service pricing.

newCreateRouteCalculator

Arguments

:: Text
$sel:calculatorName:CreateRouteCalculator'

-> Text
$sel:dataSource:CreateRouteCalculator'

-> PricingPlan
$sel:pricingPlan:CreateRouteCalculator'

-> CreateRouteCalculator

Create a value of CreateRouteCalculator with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:description:CreateRouteCalculator', createRouteCalculator_description - The optional description for the route calculator resource.

$sel:tags:CreateRouteCalculator', createRouteCalculator_tags - Applies one or more tags to the route calculator resource. A tag is a key-value pair helps manage, identify, search, and filter your resources by labelling them.

For example: { "tag1" : "value1", "tag2" : "value2"}

Format: "key" : "value"

Restrictions:

Maximum 50 tags per resource

Each resource tag must be unique with a maximum of one value.

Maximum key length: 128 Unicode characters in UTF-8

Maximum value length: 256 Unicode characters in UTF-8

Can use alphanumeric characters (A–Z, a–z, 0–9), and the following characters: + - = . _ : / @.

$sel:calculatorName:CreateRouteCalculator', createRouteCalculator_calculatorName - The name of the route calculator resource.

Requirements:

Can use alphanumeric characters (A–Z, a–z, 0–9) , hyphens (-), periods (.), and underscores (_).

Must be a unique Route calculator resource name.

No spaces allowed. For example, ExampleRouteCalculator.

$sel:dataSource:CreateRouteCalculator', createRouteCalculator_dataSource - Specifies the data provider of traffic and road network data.

This field is case-sensitive. Enter the valid values as shown. For example, entering HERE returns an error. Route calculators that use Esri as a data source only calculate routes that are shorter than 400 km.

Valid values include:

Esri – For additional information about Esri's coverage in your region of interest, see Esri details on street networks and traffic coverage.

Here – For additional information about HERE Technologies' coverage in your region of interest, see HERE car routing coverage and HERE truck routing coverage.

For additional information , see Data providers on the Amazon Location Service Developer Guide.

$sel:pricingPlan:CreateRouteCalculator', createRouteCalculator_pricingPlan - Specifies the pricing plan for your route calculator resource.

For additional details and restrictions on each pricing plan option, see Amazon Location Service pricing.

Request Lenses

createRouteCalculator_description :: Lens' CreateRouteCalculator (Maybe Text)

The optional description for the route calculator resource.

createRouteCalculator_tags :: Lens' CreateRouteCalculator (Maybe (HashMap Text Text))

Applies one or more tags to the route calculator resource. A tag is a key-value pair helps manage, identify, search, and filter your resources by labelling them.

For example: { "tag1" : "value1", "tag2" : "value2"}

Format: "key" : "value"

Restrictions:

Maximum 50 tags per resource

Each resource tag must be unique with a maximum of one value.

Maximum key length: 128 Unicode characters in UTF-8

Maximum value length: 256 Unicode characters in UTF-8

Can use alphanumeric characters (A–Z, a–z, 0–9), and the following characters: + - = . _ : / @.

createRouteCalculator_calculatorName :: Lens' CreateRouteCalculator Text

The name of the route calculator resource.

Requirements:

Can use alphanumeric characters (A–Z, a–z, 0–9) , hyphens (-), periods (.), and underscores (_).

Must be a unique Route calculator resource name.

No spaces allowed. For example, ExampleRouteCalculator.

createRouteCalculator_dataSource :: Lens' CreateRouteCalculator Text

Specifies the data provider of traffic and road network data.

This field is case-sensitive. Enter the valid values as shown. For example, entering HERE returns an error. Route calculators that use Esri as a data source only calculate routes that are shorter than 400 km.

Valid values include:

Esri – For additional information about Esri's coverage in your region of interest, see Esri details on street networks and traffic coverage.

Here – For additional information about HERE Technologies' coverage in your region of interest, see HERE car routing coverage and HERE truck routing coverage.

For additional information , see Data providers on the Amazon Location Service Developer Guide.

createRouteCalculator_pricingPlan :: Lens' CreateRouteCalculator PricingPlan

Specifies the pricing plan for your route calculator resource.

For additional details and restrictions on each pricing plan option, see Amazon Location Service pricing.

Destructuring the Response

data CreateRouteCalculatorResponse

See: newCreateRouteCalculatorResponse smart constructor.

Constructors

CreateRouteCalculatorResponse'

Fields

httpStatus :: Int

The response's http status code.

calculatorArn :: Text

The Amazon Resource Name (ARN) for the route calculator resource. Use the ARN when you specify a resource across all AWS.

Format example: arn:aws:geo:region:account-id:route-calculator/ExampleCalculator

calculatorName :: Text

The name of the route calculator resource.

For example, ExampleRouteCalculator.

createTime :: POSIX

The timestamp when the route calculator resource was created in ISO 8601 format: YYYY-MM-DDThh:mm:ss.sssZ.

For example, 2020–07-2T12:15:20.000Z+01:00

newCreateRouteCalculatorResponse

Arguments

:: Int
$sel:httpStatus:CreateRouteCalculatorResponse'

-> Text
$sel:calculatorArn:CreateRouteCalculatorResponse'

-> Text
$sel:calculatorName:CreateRouteCalculator'

-> UTCTime
$sel:createTime:CreateRouteCalculatorResponse'

-> CreateRouteCalculatorResponse

Create a value of CreateRouteCalculatorResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:httpStatus:CreateRouteCalculatorResponse', createRouteCalculatorResponse_httpStatus - The response's http status code.

$sel:calculatorArn:CreateRouteCalculatorResponse', createRouteCalculatorResponse_calculatorArn - The Amazon Resource Name (ARN) for the route calculator resource. Use the ARN when you specify a resource across all AWS.

Format example: arn:aws:geo:region:account-id:route-calculator/ExampleCalculator

$sel:calculatorName:CreateRouteCalculator', createRouteCalculatorResponse_calculatorName - The name of the route calculator resource.

For example, ExampleRouteCalculator.

$sel:createTime:CreateRouteCalculatorResponse', createRouteCalculatorResponse_createTime - The timestamp when the route calculator resource was created in ISO 8601 format: YYYY-MM-DDThh:mm:ss.sssZ.

For example, 2020–07-2T12:15:20.000Z+01:00

Response Lenses

createRouteCalculatorResponse_httpStatus :: Lens' CreateRouteCalculatorResponse Int

The response's http status code.

createRouteCalculatorResponse_calculatorArn :: Lens' CreateRouteCalculatorResponse Text

The Amazon Resource Name (ARN) for the route calculator resource. Use the ARN when you specify a resource across all AWS.

Format example: arn:aws:geo:region:account-id:route-calculator/ExampleCalculator

createRouteCalculatorResponse_calculatorName :: Lens' CreateRouteCalculatorResponse Text

The name of the route calculator resource.

For example, ExampleRouteCalculator.

createRouteCalculatorResponse_createTime :: Lens' CreateRouteCalculatorResponse UTCTime

The timestamp when the route calculator resource was created in ISO 8601 format: YYYY-MM-DDThh:mm:ss.sssZ.

For example, 2020–07-2T12:15:20.000Z+01:00