Amazonka.Location.CreateTracker

Description

Creates a tracker resource in your AWS account, which lets you retrieve current and historical location of devices.

Creating a Request

data CreateTracker

See: newCreateTracker smart constructor.

Constructors

CreateTracker'

Fields

pricingPlanDataSource :: Maybe Text

Specifies the data provider for the tracker resource.

Required value for the following pricing plans: MobileAssetTracking | MobileAssetManagement

For more information about Data Providers, and Pricing plans, see the Amazon Location Service product page.

Amazon Location Service only uses PricingPlanDataSource to calculate billing for your tracker resource. Your data will not be shared with the data provider, and will remain in your AWS account or Region unless you move it.

Valid values: Esri | Here

kmsKeyId :: Maybe Text

A key identifier for an AWS KMS customer managed key. Enter a key ID, key ARN, alias name, or alias ARN.

description :: Maybe Text

An optional description for the tracker resource.

tags :: Maybe (HashMap Text Text)

Applies one or more tags to the tracker resource. A tag is a key-value pair helps manage, identify, search, and filter your resources by labelling them.

Format: "key" : "value"

Restrictions:

Maximum 50 tags per resource

Each resource tag must be unique with a maximum of one value.

Maximum key length: 128 Unicode characters in UTF-8

Maximum value length: 256 Unicode characters in UTF-8

Can use alphanumeric characters (A–Z, a–z, 0–9), and the following characters: + - = . _ : / @.

positionFiltering :: Maybe PositionFiltering

Specifies the position filtering for the tracker resource.

Valid values:

TimeBased - Location updates are evaluated against linked geofence collections, but not every location update is stored. If your update frequency is more often than 30 seconds, only one update per 30 seconds is stored for each unique device ID.

DistanceBased - If the device has moved less than 30 m (98.4 ft), location updates are ignored. Location updates within this distance are neither evaluated against linked geofence collections, nor stored. This helps control costs by reducing the number of geofence evaluations and device positions to retrieve. Distance-based filtering can also reduce the jitter effect when displaying device trajectory on a map.

This field is optional. If not specified, the default value is TimeBased.

pricingPlan :: PricingPlan

Specifies the pricing plan for the tracker resource.

For additional details and restrictions on each pricing plan option, see Amazon Location Service pricing.

trackerName :: Text

The name for the tracker resource.

Requirements:

Contain only alphanumeric characters (A-Z, a-z, 0-9) , hyphens (-), periods (.), and underscores (_).

Must be a unique tracker resource name.

No spaces allowed. For example, ExampleTracker.

newCreateTracker

Arguments

:: PricingPlan
$sel:pricingPlan:CreateTracker'

-> Text
$sel:trackerName:CreateTracker'

-> CreateTracker

Create a value of CreateTracker with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:pricingPlanDataSource:CreateTracker', createTracker_pricingPlanDataSource - Specifies the data provider for the tracker resource.

Required value for the following pricing plans: MobileAssetTracking | MobileAssetManagement

For more information about Data Providers, and Pricing plans, see the Amazon Location Service product page.

Amazon Location Service only uses PricingPlanDataSource to calculate billing for your tracker resource. Your data will not be shared with the data provider, and will remain in your AWS account or Region unless you move it.

Valid values: Esri | Here

$sel:kmsKeyId:CreateTracker', createTracker_kmsKeyId - A key identifier for an AWS KMS customer managed key. Enter a key ID, key ARN, alias name, or alias ARN.

$sel:description:CreateTracker', createTracker_description - An optional description for the tracker resource.

$sel:tags:CreateTracker', createTracker_tags - Applies one or more tags to the tracker resource. A tag is a key-value pair helps manage, identify, search, and filter your resources by labelling them.

Format: "key" : "value"

Restrictions:

Maximum 50 tags per resource

Each resource tag must be unique with a maximum of one value.

Maximum key length: 128 Unicode characters in UTF-8

Maximum value length: 256 Unicode characters in UTF-8

Can use alphanumeric characters (A–Z, a–z, 0–9), and the following characters: + - = . _ : / @.

$sel:positionFiltering:CreateTracker', createTracker_positionFiltering - Specifies the position filtering for the tracker resource.

Valid values:

TimeBased - Location updates are evaluated against linked geofence collections, but not every location update is stored. If your update frequency is more often than 30 seconds, only one update per 30 seconds is stored for each unique device ID.

DistanceBased - If the device has moved less than 30 m (98.4 ft), location updates are ignored. Location updates within this distance are neither evaluated against linked geofence collections, nor stored. This helps control costs by reducing the number of geofence evaluations and device positions to retrieve. Distance-based filtering can also reduce the jitter effect when displaying device trajectory on a map.

This field is optional. If not specified, the default value is TimeBased.

$sel:pricingPlan:CreateTracker', createTracker_pricingPlan - Specifies the pricing plan for the tracker resource.

For additional details and restrictions on each pricing plan option, see Amazon Location Service pricing.

$sel:trackerName:CreateTracker', createTracker_trackerName - The name for the tracker resource.

Requirements:

Contain only alphanumeric characters (A-Z, a-z, 0-9) , hyphens (-), periods (.), and underscores (_).

Must be a unique tracker resource name.

No spaces allowed. For example, ExampleTracker.

Request Lenses

createTracker_pricingPlanDataSource :: Lens' CreateTracker (Maybe Text)

Specifies the data provider for the tracker resource.

Required value for the following pricing plans: MobileAssetTracking | MobileAssetManagement

For more information about Data Providers, and Pricing plans, see the Amazon Location Service product page.

Amazon Location Service only uses PricingPlanDataSource to calculate billing for your tracker resource. Your data will not be shared with the data provider, and will remain in your AWS account or Region unless you move it.

Valid values: Esri | Here

createTracker_kmsKeyId :: Lens' CreateTracker (Maybe Text)

A key identifier for an AWS KMS customer managed key. Enter a key ID, key ARN, alias name, or alias ARN.

createTracker_description :: Lens' CreateTracker (Maybe Text)

An optional description for the tracker resource.

createTracker_tags :: Lens' CreateTracker (Maybe (HashMap Text Text))

Applies one or more tags to the tracker resource. A tag is a key-value pair helps manage, identify, search, and filter your resources by labelling them.

Format: "key" : "value"

Restrictions:

Maximum 50 tags per resource

Each resource tag must be unique with a maximum of one value.

Maximum key length: 128 Unicode characters in UTF-8

Maximum value length: 256 Unicode characters in UTF-8

Can use alphanumeric characters (A–Z, a–z, 0–9), and the following characters: + - = . _ : / @.

createTracker_positionFiltering :: Lens' CreateTracker (Maybe PositionFiltering)

Specifies the position filtering for the tracker resource.

Valid values:

TimeBased - Location updates are evaluated against linked geofence collections, but not every location update is stored. If your update frequency is more often than 30 seconds, only one update per 30 seconds is stored for each unique device ID.

DistanceBased - If the device has moved less than 30 m (98.4 ft), location updates are ignored. Location updates within this distance are neither evaluated against linked geofence collections, nor stored. This helps control costs by reducing the number of geofence evaluations and device positions to retrieve. Distance-based filtering can also reduce the jitter effect when displaying device trajectory on a map.

This field is optional. If not specified, the default value is TimeBased.

createTracker_pricingPlan :: Lens' CreateTracker PricingPlan

Specifies the pricing plan for the tracker resource.

For additional details and restrictions on each pricing plan option, see Amazon Location Service pricing.

createTracker_trackerName :: Lens' CreateTracker Text

The name for the tracker resource.

Requirements:

Contain only alphanumeric characters (A-Z, a-z, 0-9) , hyphens (-), periods (.), and underscores (_).

Must be a unique tracker resource name.

No spaces allowed. For example, ExampleTracker.

Destructuring the Response

data CreateTrackerResponse

See: newCreateTrackerResponse smart constructor.

Constructors

CreateTrackerResponse'

Fields

httpStatus :: Int

The response's http status code.

createTime :: POSIX

The timestamp for when the tracker resource was created in ISO 8601 format: YYYY-MM-DDThh:mm:ss.sssZ.

trackerArn :: Text

The Amazon Resource Name (ARN) for the tracker resource. Used when you need to specify a resource across all AWS.

Format example: arn:aws:geo:region:account-id:tracker/ExampleTracker

trackerName :: Text

The name of the tracker resource.

newCreateTrackerResponse

Arguments

:: Int
$sel:httpStatus:CreateTrackerResponse'

-> UTCTime
$sel:createTime:CreateTrackerResponse'

-> Text
$sel:trackerArn:CreateTrackerResponse'

-> Text
$sel:trackerName:CreateTracker'

-> CreateTrackerResponse

Create a value of CreateTrackerResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:httpStatus:CreateTrackerResponse', createTrackerResponse_httpStatus - The response's http status code.

$sel:createTime:CreateTrackerResponse', createTrackerResponse_createTime - The timestamp for when the tracker resource was created in ISO 8601 format: YYYY-MM-DDThh:mm:ss.sssZ.

$sel:trackerArn:CreateTrackerResponse', createTrackerResponse_trackerArn - The Amazon Resource Name (ARN) for the tracker resource. Used when you need to specify a resource across all AWS.

Format example: arn:aws:geo:region:account-id:tracker/ExampleTracker

$sel:trackerName:CreateTracker', createTrackerResponse_trackerName - The name of the tracker resource.

Response Lenses

createTrackerResponse_httpStatus :: Lens' CreateTrackerResponse Int

The response's http status code.

createTrackerResponse_createTime :: Lens' CreateTrackerResponse UTCTime

The timestamp for when the tracker resource was created in ISO 8601 format: YYYY-MM-DDThh:mm:ss.sssZ.

createTrackerResponse_trackerArn :: Lens' CreateTrackerResponse Text

The Amazon Resource Name (ARN) for the tracker resource. Used when you need to specify a resource across all AWS.

Format example: arn:aws:geo:region:account-id:tracker/ExampleTracker

createTrackerResponse_trackerName :: Lens' CreateTrackerResponse Text

The name of the tracker resource.