Amazonka.Location.DescribePlaceIndex

Description

Retrieves the place index resource details.

Creating a Request

data DescribePlaceIndex

See: newDescribePlaceIndex smart constructor.

Constructors

DescribePlaceIndex'

Fields

indexName :: Text

The name of the place index resource.

newDescribePlaceIndex

Arguments

:: Text
$sel:indexName:DescribePlaceIndex'

-> DescribePlaceIndex

Create a value of DescribePlaceIndex with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:indexName:DescribePlaceIndex', describePlaceIndex_indexName - The name of the place index resource.

Request Lenses

describePlaceIndex_indexName :: Lens' DescribePlaceIndex Text

The name of the place index resource.

Destructuring the Response

data DescribePlaceIndexResponse

See: newDescribePlaceIndexResponse smart constructor.

Constructors

DescribePlaceIndexResponse'

Fields

tags :: Maybe (HashMap Text Text)

Tags associated with place index resource.

httpStatus :: Int

The response's http status code.

createTime :: POSIX

The timestamp for when the place index resource was created in ISO 8601 format: YYYY-MM-DDThh:mm:ss.sssZ.

dataSource :: Text

The data provider of geospatial data. Indicates one of the available providers:

Esri

Here

For additional details on data providers, see Amazon Location Service data providers.

dataSourceConfiguration :: DataSourceConfiguration

The specified data storage option for requesting Places.

description :: Text

The optional description for the place index resource.

indexArn :: Text

The Amazon Resource Name (ARN) for the place index resource. Used to specify a resource across AWS.

Format example: arn:aws:geo:region:account-id:place-index/ExamplePlaceIndex

indexName :: Text

The name of the place index resource being described.

pricingPlan :: PricingPlan

The pricing plan selected for the specified place index resource.

For additional details and restrictions on each pricing plan option, see Amazon Location Service pricing.

updateTime :: POSIX

The timestamp for when the place index resource was last updated in ISO 8601 format: YYYY-MM-DDThh:mm:ss.sssZ.

newDescribePlaceIndexResponse

Arguments

:: Int
$sel:httpStatus:DescribePlaceIndexResponse'

-> UTCTime
$sel:createTime:DescribePlaceIndexResponse'

-> Text
$sel:dataSource:DescribePlaceIndexResponse'

-> DataSourceConfiguration
$sel:dataSourceConfiguration:DescribePlaceIndexResponse'

-> Text
$sel:description:DescribePlaceIndexResponse'

-> Text
$sel:indexArn:DescribePlaceIndexResponse'

-> Text
$sel:indexName:DescribePlaceIndex'

-> PricingPlan
$sel:pricingPlan:DescribePlaceIndexResponse'

-> UTCTime
$sel:updateTime:DescribePlaceIndexResponse'

-> DescribePlaceIndexResponse

Create a value of DescribePlaceIndexResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:tags:DescribePlaceIndexResponse', describePlaceIndexResponse_tags - Tags associated with place index resource.

$sel:httpStatus:DescribePlaceIndexResponse', describePlaceIndexResponse_httpStatus - The response's http status code.

$sel:createTime:DescribePlaceIndexResponse', describePlaceIndexResponse_createTime - The timestamp for when the place index resource was created in ISO 8601 format: YYYY-MM-DDThh:mm:ss.sssZ.

$sel:dataSource:DescribePlaceIndexResponse', describePlaceIndexResponse_dataSource - The data provider of geospatial data. Indicates one of the available providers:

Esri

Here

For additional details on data providers, see Amazon Location Service data providers.

$sel:dataSourceConfiguration:DescribePlaceIndexResponse', describePlaceIndexResponse_dataSourceConfiguration - The specified data storage option for requesting Places.

$sel:description:DescribePlaceIndexResponse', describePlaceIndexResponse_description - The optional description for the place index resource.

$sel:indexArn:DescribePlaceIndexResponse', describePlaceIndexResponse_indexArn - The Amazon Resource Name (ARN) for the place index resource. Used to specify a resource across AWS.

Format example: arn:aws:geo:region:account-id:place-index/ExamplePlaceIndex

$sel:indexName:DescribePlaceIndex', describePlaceIndexResponse_indexName - The name of the place index resource being described.

$sel:pricingPlan:DescribePlaceIndexResponse', describePlaceIndexResponse_pricingPlan - The pricing plan selected for the specified place index resource.

For additional details and restrictions on each pricing plan option, see Amazon Location Service pricing.

$sel:updateTime:DescribePlaceIndexResponse', describePlaceIndexResponse_updateTime - The timestamp for when the place index resource was last updated in ISO 8601 format: YYYY-MM-DDThh:mm:ss.sssZ.

Response Lenses

describePlaceIndexResponse_tags :: Lens' DescribePlaceIndexResponse (Maybe (HashMap Text Text))

Tags associated with place index resource.

describePlaceIndexResponse_httpStatus :: Lens' DescribePlaceIndexResponse Int

The response's http status code.

describePlaceIndexResponse_createTime :: Lens' DescribePlaceIndexResponse UTCTime

The timestamp for when the place index resource was created in ISO 8601 format: YYYY-MM-DDThh:mm:ss.sssZ.

describePlaceIndexResponse_dataSource :: Lens' DescribePlaceIndexResponse Text

The data provider of geospatial data. Indicates one of the available providers:

Esri

Here

For additional details on data providers, see Amazon Location Service data providers.

describePlaceIndexResponse_dataSourceConfiguration :: Lens' DescribePlaceIndexResponse DataSourceConfiguration

The specified data storage option for requesting Places.

describePlaceIndexResponse_description :: Lens' DescribePlaceIndexResponse Text

The optional description for the place index resource.

describePlaceIndexResponse_indexArn :: Lens' DescribePlaceIndexResponse Text

The Amazon Resource Name (ARN) for the place index resource. Used to specify a resource across AWS.

Format example: arn:aws:geo:region:account-id:place-index/ExamplePlaceIndex

describePlaceIndexResponse_indexName :: Lens' DescribePlaceIndexResponse Text

The name of the place index resource being described.

describePlaceIndexResponse_pricingPlan :: Lens' DescribePlaceIndexResponse PricingPlan

The pricing plan selected for the specified place index resource.

For additional details and restrictions on each pricing plan option, see Amazon Location Service pricing.

describePlaceIndexResponse_updateTime :: Lens' DescribePlaceIndexResponse UTCTime

The timestamp for when the place index resource was last updated in ISO 8601 format: YYYY-MM-DDThh:mm:ss.sssZ.