Amazonka.Location.DisassociateTrackerConsumer

Description

Removes the association between a tracker resource and a geofence collection.

Once you unlink a tracker resource from a geofence collection, the tracker positions will no longer be automatically evaluated against geofences.

Creating a Request

data DisassociateTrackerConsumer

See: newDisassociateTrackerConsumer smart constructor.

Constructors

DisassociateTrackerConsumer'

Fields

consumerArn :: Text

The Amazon Resource Name (ARN) for the geofence collection to be disassociated from the tracker resource. Used when you need to specify a resource across all AWS.

Format example: arn:aws:geo:region:account-id:geofence-collection/ExampleGeofenceCollectionConsumer

trackerName :: Text

The name of the tracker resource to be dissociated from the consumer.

newDisassociateTrackerConsumer

Arguments

:: Text
$sel:consumerArn:DisassociateTrackerConsumer'

-> Text
$sel:trackerName:DisassociateTrackerConsumer'

-> DisassociateTrackerConsumer

Create a value of DisassociateTrackerConsumer with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:consumerArn:DisassociateTrackerConsumer', disassociateTrackerConsumer_consumerArn - The Amazon Resource Name (ARN) for the geofence collection to be disassociated from the tracker resource. Used when you need to specify a resource across all AWS.

Format example: arn:aws:geo:region:account-id:geofence-collection/ExampleGeofenceCollectionConsumer

$sel:trackerName:DisassociateTrackerConsumer', disassociateTrackerConsumer_trackerName - The name of the tracker resource to be dissociated from the consumer.

Request Lenses

disassociateTrackerConsumer_consumerArn :: Lens' DisassociateTrackerConsumer Text

The Amazon Resource Name (ARN) for the geofence collection to be disassociated from the tracker resource. Used when you need to specify a resource across all AWS.

Format example: arn:aws:geo:region:account-id:geofence-collection/ExampleGeofenceCollectionConsumer

disassociateTrackerConsumer_trackerName :: Lens' DisassociateTrackerConsumer Text

The name of the tracker resource to be dissociated from the consumer.

Destructuring the Response

data DisassociateTrackerConsumerResponse

See: newDisassociateTrackerConsumerResponse smart constructor.

Constructors

DisassociateTrackerConsumerResponse'

Fields

httpStatus :: Int

The response's http status code.

newDisassociateTrackerConsumerResponse

Arguments

:: Int
$sel:httpStatus:DisassociateTrackerConsumerResponse'

-> DisassociateTrackerConsumerResponse

Create a value of DisassociateTrackerConsumerResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:httpStatus:DisassociateTrackerConsumerResponse', disassociateTrackerConsumerResponse_httpStatus - The response's http status code.

Response Lenses

disassociateTrackerConsumerResponse_httpStatus :: Lens' DisassociateTrackerConsumerResponse Int

The response's http status code.