Amazonka.Location.GetDevicePosition

Description

Retrieves a device's most recent position according to its sample time.

Device positions are deleted after 30 days.

Creating a Request

data GetDevicePosition

See: newGetDevicePosition smart constructor.

Constructors

GetDevicePosition'

Fields

deviceId :: Text

The device whose position you want to retrieve.

trackerName :: Text

The tracker resource receiving the position update.

newGetDevicePosition

Arguments

:: Text
$sel:deviceId:GetDevicePosition'

-> Text
$sel:trackerName:GetDevicePosition'

-> GetDevicePosition

Create a value of GetDevicePosition with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:deviceId:GetDevicePosition', getDevicePosition_deviceId - The device whose position you want to retrieve.

$sel:trackerName:GetDevicePosition', getDevicePosition_trackerName - The tracker resource receiving the position update.

Request Lenses

getDevicePosition_deviceId :: Lens' GetDevicePosition Text

The device whose position you want to retrieve.

getDevicePosition_trackerName :: Lens' GetDevicePosition Text

The tracker resource receiving the position update.

Destructuring the Response

data GetDevicePositionResponse

See: newGetDevicePositionResponse smart constructor.

Constructors

GetDevicePositionResponse'

Fields

deviceId :: Maybe Text

The device whose position you retrieved.

httpStatus :: Int

The response's http status code.

position :: Sensitive (NonEmpty Double)

The last known device position.

receivedTime :: POSIX

The timestamp for when the tracker resource received the device position in ISO 8601 format: YYYY-MM-DDThh:mm:ss.sssZ.

sampleTime :: POSIX

The timestamp at which the device's position was determined. Uses ISO 8601 format: YYYY-MM-DDThh:mm:ss.sssZ.

newGetDevicePositionResponse

Arguments

:: Int
$sel:httpStatus:GetDevicePositionResponse'

-> NonEmpty Double
$sel:position:GetDevicePositionResponse'

-> UTCTime
$sel:receivedTime:GetDevicePositionResponse'

-> UTCTime
$sel:sampleTime:GetDevicePositionResponse'

-> GetDevicePositionResponse

Create a value of GetDevicePositionResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:deviceId:GetDevicePosition', getDevicePositionResponse_deviceId - The device whose position you retrieved.

$sel:httpStatus:GetDevicePositionResponse', getDevicePositionResponse_httpStatus - The response's http status code.

$sel:position:GetDevicePositionResponse', getDevicePositionResponse_position - The last known device position.

$sel:receivedTime:GetDevicePositionResponse', getDevicePositionResponse_receivedTime - The timestamp for when the tracker resource received the device position in ISO 8601 format: YYYY-MM-DDThh:mm:ss.sssZ.

$sel:sampleTime:GetDevicePositionResponse', getDevicePositionResponse_sampleTime - The timestamp at which the device's position was determined. Uses ISO 8601 format: YYYY-MM-DDThh:mm:ss.sssZ.

Response Lenses

getDevicePositionResponse_deviceId :: Lens' GetDevicePositionResponse (Maybe Text)

The device whose position you retrieved.

getDevicePositionResponse_httpStatus :: Lens' GetDevicePositionResponse Int

The response's http status code.

getDevicePositionResponse_position :: Lens' GetDevicePositionResponse (NonEmpty Double)

The last known device position.

getDevicePositionResponse_receivedTime :: Lens' GetDevicePositionResponse UTCTime

The timestamp for when the tracker resource received the device position in ISO 8601 format: YYYY-MM-DDThh:mm:ss.sssZ.

getDevicePositionResponse_sampleTime :: Lens' GetDevicePositionResponse UTCTime

The timestamp at which the device's position was determined. Uses ISO 8601 format: YYYY-MM-DDThh:mm:ss.sssZ.