Amazonka.Location.GetDevicePositionHistory

Description

Retrieves the device position history from a tracker resource within a specified range of time.

Device positions are deleted after 30 days.

This operation returns paginated results.

Creating a Request

data GetDevicePositionHistory

See: newGetDevicePositionHistory smart constructor.

Constructors

GetDevicePositionHistory'

Fields

nextToken :: Maybe Text

The pagination token specifying which page of results to return in the response. If no token is provided, the default page is the first page.

Default value: null

endTimeExclusive :: Maybe POSIX

Specify the end time for the position history in ISO 8601 format: YYYY-MM-DDThh:mm:ss.sssZ. By default, the value will be the time that the request is made.

Requirement:

The time specified for EndTimeExclusive must be after the time for StartTimeInclusive.

startTimeInclusive :: Maybe POSIX

Specify the start time for the position history in ISO 8601 format: YYYY-MM-DDThh:mm:ss.sssZ. By default, the value will be 24 hours prior to the time that the request is made.

Requirement:

The time specified for StartTimeInclusive must be before EndTimeExclusive.

deviceId :: Text

The device whose position history you want to retrieve.

trackerName :: Text

The tracker resource receiving the request for the device position history.

newGetDevicePositionHistory

Arguments

:: Text
$sel:deviceId:GetDevicePositionHistory'

-> Text
$sel:trackerName:GetDevicePositionHistory'

-> GetDevicePositionHistory

Create a value of GetDevicePositionHistory with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:nextToken:GetDevicePositionHistory', getDevicePositionHistory_nextToken - The pagination token specifying which page of results to return in the response. If no token is provided, the default page is the first page.

Default value: null

$sel:endTimeExclusive:GetDevicePositionHistory', getDevicePositionHistory_endTimeExclusive - Specify the end time for the position history in ISO 8601 format: YYYY-MM-DDThh:mm:ss.sssZ. By default, the value will be the time that the request is made.

Requirement:

The time specified for EndTimeExclusive must be after the time for StartTimeInclusive.

$sel:startTimeInclusive:GetDevicePositionHistory', getDevicePositionHistory_startTimeInclusive - Specify the start time for the position history in ISO 8601 format: YYYY-MM-DDThh:mm:ss.sssZ. By default, the value will be 24 hours prior to the time that the request is made.

Requirement:

The time specified for StartTimeInclusive must be before EndTimeExclusive.

$sel:deviceId:GetDevicePositionHistory', getDevicePositionHistory_deviceId - The device whose position history you want to retrieve.

$sel:trackerName:GetDevicePositionHistory', getDevicePositionHistory_trackerName - The tracker resource receiving the request for the device position history.

Request Lenses

getDevicePositionHistory_nextToken :: Lens' GetDevicePositionHistory (Maybe Text)

The pagination token specifying which page of results to return in the response. If no token is provided, the default page is the first page.

Default value: null

getDevicePositionHistory_endTimeExclusive :: Lens' GetDevicePositionHistory (Maybe UTCTime)

Specify the end time for the position history in ISO 8601 format: YYYY-MM-DDThh:mm:ss.sssZ. By default, the value will be the time that the request is made.

Requirement:

The time specified for EndTimeExclusive must be after the time for StartTimeInclusive.

getDevicePositionHistory_startTimeInclusive :: Lens' GetDevicePositionHistory (Maybe UTCTime)

Specify the start time for the position history in ISO 8601 format: YYYY-MM-DDThh:mm:ss.sssZ. By default, the value will be 24 hours prior to the time that the request is made.

Requirement:

The time specified for StartTimeInclusive must be before EndTimeExclusive.

getDevicePositionHistory_deviceId :: Lens' GetDevicePositionHistory Text

The device whose position history you want to retrieve.

getDevicePositionHistory_trackerName :: Lens' GetDevicePositionHistory Text

The tracker resource receiving the request for the device position history.

Destructuring the Response

data GetDevicePositionHistoryResponse

See: newGetDevicePositionHistoryResponse smart constructor.

Constructors

GetDevicePositionHistoryResponse'

Fields

nextToken :: Maybe Text

A pagination token indicating there are additional pages available. You can use the token in a following request to fetch the next set of results.

httpStatus :: Int

The response's http status code.

devicePositions :: [DevicePosition]

Contains the position history details for the requested device.

newGetDevicePositionHistoryResponse

Arguments

:: Int
$sel:httpStatus:GetDevicePositionHistoryResponse'

-> GetDevicePositionHistoryResponse

Create a value of GetDevicePositionHistoryResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:nextToken:GetDevicePositionHistory', getDevicePositionHistoryResponse_nextToken - A pagination token indicating there are additional pages available. You can use the token in a following request to fetch the next set of results.

$sel:httpStatus:GetDevicePositionHistoryResponse', getDevicePositionHistoryResponse_httpStatus - The response's http status code.

$sel:devicePositions:GetDevicePositionHistoryResponse', getDevicePositionHistoryResponse_devicePositions - Contains the position history details for the requested device.

Response Lenses

getDevicePositionHistoryResponse_nextToken :: Lens' GetDevicePositionHistoryResponse (Maybe Text)

A pagination token indicating there are additional pages available. You can use the token in a following request to fetch the next set of results.

getDevicePositionHistoryResponse_httpStatus :: Lens' GetDevicePositionHistoryResponse Int

The response's http status code.

getDevicePositionHistoryResponse_devicePositions :: Lens' GetDevicePositionHistoryResponse [DevicePosition]

Contains the position history details for the requested device.