Amazonka.Location.GetGeofence

Description

Retrieves the geofence details from a geofence collection.

Creating a Request

data GetGeofence

See: newGetGeofence smart constructor.

Constructors

GetGeofence'

Fields

collectionName :: Text

The geofence collection storing the target geofence.

geofenceId :: Text

The geofence you're retrieving details for.

newGetGeofence

Arguments

:: Text
$sel:collectionName:GetGeofence'

-> Text
$sel:geofenceId:GetGeofence'

-> GetGeofence

Create a value of GetGeofence with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:collectionName:GetGeofence', getGeofence_collectionName - The geofence collection storing the target geofence.

$sel:geofenceId:GetGeofence', getGeofence_geofenceId - The geofence you're retrieving details for.

Request Lenses

getGeofence_collectionName :: Lens' GetGeofence Text

The geofence collection storing the target geofence.

getGeofence_geofenceId :: Lens' GetGeofence Text

The geofence you're retrieving details for.

Destructuring the Response

data GetGeofenceResponse

See: newGetGeofenceResponse smart constructor.

Constructors

GetGeofenceResponse'

Fields

httpStatus :: Int

The response's http status code.

createTime :: POSIX

The timestamp for when the geofence collection was created in ISO 8601 format: YYYY-MM-DDThh:mm:ss.sssZ

geofenceId :: Text

The geofence identifier.

geometry :: GeofenceGeometry

Contains the geofence geometry details describing a polygon.

status :: Text

Identifies the state of the geofence. A geofence will hold one of the following states:

ACTIVE — The geofence has been indexed by the system.

PENDING — The geofence is being processed by the system.

FAILED — The geofence failed to be indexed by the system.

DELETED — The geofence has been deleted from the system index.

DELETING — The geofence is being deleted from the system index.

updateTime :: POSIX

The timestamp for when the geofence collection was last updated in ISO 8601 format: YYYY-MM-DDThh:mm:ss.sssZ

newGetGeofenceResponse

Arguments

:: Int
$sel:httpStatus:GetGeofenceResponse'

-> UTCTime
$sel:createTime:GetGeofenceResponse'

-> Text
$sel:geofenceId:GetGeofence'

-> GeofenceGeometry
$sel:geometry:GetGeofenceResponse'

-> Text
$sel:status:GetGeofenceResponse'

-> UTCTime
$sel:updateTime:GetGeofenceResponse'

-> GetGeofenceResponse

Create a value of GetGeofenceResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:httpStatus:GetGeofenceResponse', getGeofenceResponse_httpStatus - The response's http status code.

$sel:createTime:GetGeofenceResponse', getGeofenceResponse_createTime - The timestamp for when the geofence collection was created in ISO 8601 format: YYYY-MM-DDThh:mm:ss.sssZ

$sel:geofenceId:GetGeofence', getGeofenceResponse_geofenceId - The geofence identifier.

$sel:geometry:GetGeofenceResponse', getGeofenceResponse_geometry - Contains the geofence geometry details describing a polygon.

$sel:status:GetGeofenceResponse', getGeofenceResponse_status - Identifies the state of the geofence. A geofence will hold one of the following states:

ACTIVE — The geofence has been indexed by the system.

PENDING — The geofence is being processed by the system.

FAILED — The geofence failed to be indexed by the system.

DELETED — The geofence has been deleted from the system index.

DELETING — The geofence is being deleted from the system index.

$sel:updateTime:GetGeofenceResponse', getGeofenceResponse_updateTime - The timestamp for when the geofence collection was last updated in ISO 8601 format: YYYY-MM-DDThh:mm:ss.sssZ

Response Lenses

getGeofenceResponse_httpStatus :: Lens' GetGeofenceResponse Int

The response's http status code.

getGeofenceResponse_createTime :: Lens' GetGeofenceResponse UTCTime

The timestamp for when the geofence collection was created in ISO 8601 format: YYYY-MM-DDThh:mm:ss.sssZ

getGeofenceResponse_geofenceId :: Lens' GetGeofenceResponse Text

The geofence identifier.

getGeofenceResponse_geometry :: Lens' GetGeofenceResponse GeofenceGeometry

Contains the geofence geometry details describing a polygon.

getGeofenceResponse_status :: Lens' GetGeofenceResponse Text

Identifies the state of the geofence. A geofence will hold one of the following states:

ACTIVE — The geofence has been indexed by the system.

PENDING — The geofence is being processed by the system.

FAILED — The geofence failed to be indexed by the system.

DELETED — The geofence has been deleted from the system index.

DELETING — The geofence is being deleted from the system index.

getGeofenceResponse_updateTime :: Lens' GetGeofenceResponse UTCTime

The timestamp for when the geofence collection was last updated in ISO 8601 format: YYYY-MM-DDThh:mm:ss.sssZ