Amazonka.Location.GetMapGlyphs

Description

Retrieves glyphs used to display labels on a map.

Creating a Request

data GetMapGlyphs

See: newGetMapGlyphs smart constructor.

Constructors

GetMapGlyphs'

Fields

fontStack :: Text

A comma-separated list of fonts to load glyphs from in order of preference. For example, Noto Sans Regular, Arial Unicode.

Valid fonts stacks for Esri styles:

VectorEsriDarkGrayCanvas – Ubuntu Medium Italic | Ubuntu Medium | Ubuntu Italic | Ubuntu Regular | Ubuntu Bold

VectorEsriLightGrayCanvas – Ubuntu Italic | Ubuntu Regular | Ubuntu Light | Ubuntu Bold

VectorEsriTopographic – Noto Sans Italic | Noto Sans Regular | Noto Sans Bold | Noto Serif Regular | Roboto Condensed Light Italic

VectorEsriStreets – Arial Regular | Arial Italic | Arial Bold

VectorEsriNavigation – Arial Regular | Arial Italic | Arial Bold

Valid font stacks for HERE Technologies styles:

VectorHereBerlin – Fira GO Regular | Fira GO Bold

fontUnicodeRange :: Text

A Unicode range of characters to download glyphs for. Each response will contain 256 characters. For example, 0–255 includes all characters from range U+0000 to 00FF. Must be aligned to multiples of 256.

mapName :: Text

The map resource associated with the glyph ﬁle.

newGetMapGlyphs

Arguments

:: Text
$sel:fontStack:GetMapGlyphs'

-> Text
$sel:fontUnicodeRange:GetMapGlyphs'

-> Text
$sel:mapName:GetMapGlyphs'

-> GetMapGlyphs

Create a value of GetMapGlyphs with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:fontStack:GetMapGlyphs', getMapGlyphs_fontStack - A comma-separated list of fonts to load glyphs from in order of preference. For example, Noto Sans Regular, Arial Unicode.

Valid fonts stacks for Esri styles:

VectorEsriDarkGrayCanvas – Ubuntu Medium Italic | Ubuntu Medium | Ubuntu Italic | Ubuntu Regular | Ubuntu Bold

VectorEsriLightGrayCanvas – Ubuntu Italic | Ubuntu Regular | Ubuntu Light | Ubuntu Bold

VectorEsriTopographic – Noto Sans Italic | Noto Sans Regular | Noto Sans Bold | Noto Serif Regular | Roboto Condensed Light Italic

VectorEsriStreets – Arial Regular | Arial Italic | Arial Bold

VectorEsriNavigation – Arial Regular | Arial Italic | Arial Bold

Valid font stacks for HERE Technologies styles:

VectorHereBerlin – Fira GO Regular | Fira GO Bold

$sel:fontUnicodeRange:GetMapGlyphs', getMapGlyphs_fontUnicodeRange - A Unicode range of characters to download glyphs for. Each response will contain 256 characters. For example, 0–255 includes all characters from range U+0000 to 00FF. Must be aligned to multiples of 256.

$sel:mapName:GetMapGlyphs', getMapGlyphs_mapName - The map resource associated with the glyph ﬁle.

Request Lenses

getMapGlyphs_fontStack :: Lens' GetMapGlyphs Text

A comma-separated list of fonts to load glyphs from in order of preference. For example, Noto Sans Regular, Arial Unicode.

Valid fonts stacks for Esri styles:

VectorEsriDarkGrayCanvas – Ubuntu Medium Italic | Ubuntu Medium | Ubuntu Italic | Ubuntu Regular | Ubuntu Bold

VectorEsriLightGrayCanvas – Ubuntu Italic | Ubuntu Regular | Ubuntu Light | Ubuntu Bold

VectorEsriTopographic – Noto Sans Italic | Noto Sans Regular | Noto Sans Bold | Noto Serif Regular | Roboto Condensed Light Italic

VectorEsriStreets – Arial Regular | Arial Italic | Arial Bold

VectorEsriNavigation – Arial Regular | Arial Italic | Arial Bold

Valid font stacks for HERE Technologies styles:

VectorHereBerlin – Fira GO Regular | Fira GO Bold

getMapGlyphs_fontUnicodeRange :: Lens' GetMapGlyphs Text

A Unicode range of characters to download glyphs for. Each response will contain 256 characters. For example, 0–255 includes all characters from range U+0000 to 00FF. Must be aligned to multiples of 256.

getMapGlyphs_mapName :: Lens' GetMapGlyphs Text

The map resource associated with the glyph ﬁle.

Destructuring the Response

data GetMapGlyphsResponse

See: newGetMapGlyphsResponse smart constructor.

Constructors

GetMapGlyphsResponse'

Fields

blob :: Maybe ByteString

The blob's content type.

contentType :: Maybe Text

The map glyph content type. For example, application/octet-stream.

httpStatus :: Int

The response's http status code.

newGetMapGlyphsResponse

Arguments

:: Int
$sel:httpStatus:GetMapGlyphsResponse'

-> GetMapGlyphsResponse

Create a value of GetMapGlyphsResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:blob:GetMapGlyphsResponse', getMapGlyphsResponse_blob - The blob's content type.

$sel:contentType:GetMapGlyphsResponse', getMapGlyphsResponse_contentType - The map glyph content type. For example, application/octet-stream.

$sel:httpStatus:GetMapGlyphsResponse', getMapGlyphsResponse_httpStatus - The response's http status code.

Response Lenses

getMapGlyphsResponse_blob :: Lens' GetMapGlyphsResponse (Maybe ByteString)

The blob's content type.

getMapGlyphsResponse_contentType :: Lens' GetMapGlyphsResponse (Maybe Text)

The map glyph content type. For example, application/octet-stream.

getMapGlyphsResponse_httpStatus :: Lens' GetMapGlyphsResponse Int

The response's http status code.