Amazonka.Location.ListMaps

Description

Lists map resources in your AWS account.

This operation returns paginated results.

Creating a Request

data ListMaps

See: newListMaps smart constructor.

Constructors

ListMaps'

Fields

nextToken :: Maybe Text

The pagination token specifying which page of results to return in the response. If no token is provided, the default page is the first page.

Default value: null

maxResults :: Maybe Natural

An optional limit for the number of resources returned in a single call.

Default value: 100

newListMaps :: ListMaps

Create a value of ListMaps with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:nextToken:ListMaps', listMaps_nextToken - The pagination token specifying which page of results to return in the response. If no token is provided, the default page is the first page.

Default value: null

$sel:maxResults:ListMaps', listMaps_maxResults - An optional limit for the number of resources returned in a single call.

Default value: 100

Request Lenses

listMaps_nextToken :: Lens' ListMaps (Maybe Text)

The pagination token specifying which page of results to return in the response. If no token is provided, the default page is the first page.

Default value: null

listMaps_maxResults :: Lens' ListMaps (Maybe Natural)

An optional limit for the number of resources returned in a single call.

Default value: 100

Destructuring the Response

data ListMapsResponse

See: newListMapsResponse smart constructor.

Constructors

ListMapsResponse'

Fields

nextToken :: Maybe Text

A pagination token indicating there are additional pages available. You can use the token in a following request to fetch the next set of results.

httpStatus :: Int

The response's http status code.

entries :: [ListMapsResponseEntry]

Contains a list of maps in your AWS account

newListMapsResponse

Arguments

:: Int
$sel:httpStatus:ListMapsResponse'

-> ListMapsResponse

Create a value of ListMapsResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:nextToken:ListMaps', listMapsResponse_nextToken - A pagination token indicating there are additional pages available. You can use the token in a following request to fetch the next set of results.

$sel:httpStatus:ListMapsResponse', listMapsResponse_httpStatus - The response's http status code.

$sel:entries:ListMapsResponse', listMapsResponse_entries - Contains a list of maps in your AWS account

Response Lenses

listMapsResponse_nextToken :: Lens' ListMapsResponse (Maybe Text)

A pagination token indicating there are additional pages available. You can use the token in a following request to fetch the next set of results.

listMapsResponse_httpStatus :: Lens' ListMapsResponse Int

The response's http status code.

listMapsResponse_entries :: Lens' ListMapsResponse [ListMapsResponseEntry]

Contains a list of maps in your AWS account