Amazonka.Location.ListTrackerConsumers

Description

Lists geofence collections currently associated to the given tracker resource.

This operation returns paginated results.

Creating a Request

data ListTrackerConsumers

See: newListTrackerConsumers smart constructor.

Constructors

ListTrackerConsumers'

Fields

nextToken :: Maybe Text

The pagination token specifying which page of results to return in the response. If no token is provided, the default page is the first page.

Default value: null

maxResults :: Maybe Natural

An optional limit for the number of resources returned in a single call.

Default value: 100

trackerName :: Text

The tracker resource whose associated geofence collections you want to list.

newListTrackerConsumers

Arguments

:: Text
$sel:trackerName:ListTrackerConsumers'

-> ListTrackerConsumers

Create a value of ListTrackerConsumers with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:nextToken:ListTrackerConsumers', listTrackerConsumers_nextToken - The pagination token specifying which page of results to return in the response. If no token is provided, the default page is the first page.

Default value: null

$sel:maxResults:ListTrackerConsumers', listTrackerConsumers_maxResults - An optional limit for the number of resources returned in a single call.

Default value: 100

$sel:trackerName:ListTrackerConsumers', listTrackerConsumers_trackerName - The tracker resource whose associated geofence collections you want to list.

Request Lenses

listTrackerConsumers_nextToken :: Lens' ListTrackerConsumers (Maybe Text)

The pagination token specifying which page of results to return in the response. If no token is provided, the default page is the first page.

Default value: null

listTrackerConsumers_maxResults :: Lens' ListTrackerConsumers (Maybe Natural)

An optional limit for the number of resources returned in a single call.

Default value: 100

listTrackerConsumers_trackerName :: Lens' ListTrackerConsumers Text

The tracker resource whose associated geofence collections you want to list.

Destructuring the Response

data ListTrackerConsumersResponse

See: newListTrackerConsumersResponse smart constructor.

Constructors

ListTrackerConsumersResponse'

Fields

nextToken :: Maybe Text

A pagination token indicating there are additional pages available. You can use the token in a following request to fetch the next set of results.

httpStatus :: Int

The response's http status code.

consumerArns :: [Text]

Contains the list of geofence collection ARNs associated to the tracker resource.

newListTrackerConsumersResponse

Arguments

:: Int
$sel:httpStatus:ListTrackerConsumersResponse'

-> ListTrackerConsumersResponse

Create a value of ListTrackerConsumersResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:nextToken:ListTrackerConsumers', listTrackerConsumersResponse_nextToken - A pagination token indicating there are additional pages available. You can use the token in a following request to fetch the next set of results.

$sel:httpStatus:ListTrackerConsumersResponse', listTrackerConsumersResponse_httpStatus - The response's http status code.

$sel:consumerArns:ListTrackerConsumersResponse', listTrackerConsumersResponse_consumerArns - Contains the list of geofence collection ARNs associated to the tracker resource.

Response Lenses

listTrackerConsumersResponse_nextToken :: Lens' ListTrackerConsumersResponse (Maybe Text)

A pagination token indicating there are additional pages available. You can use the token in a following request to fetch the next set of results.

listTrackerConsumersResponse_httpStatus :: Lens' ListTrackerConsumersResponse Int

The response's http status code.

listTrackerConsumersResponse_consumerArns :: Lens' ListTrackerConsumersResponse [Text]

Contains the list of geofence collection ARNs associated to the tracker resource.