Amazonka.Location.SearchPlaceIndexForPosition

Description

Reverse geocodes a given coordinate and returns a legible address. Allows you to search for Places or points of interest near a given position.

Creating a Request

data SearchPlaceIndexForPosition

See: newSearchPlaceIndexForPosition smart constructor.

Constructors

SearchPlaceIndexForPosition'

Fields

maxResults :: Maybe Natural

An optional paramer. The maximum number of results returned per request.

Default value: 50

indexName :: Text

The name of the place index resource you want to use for the search.

position :: Sensitive (NonEmpty Double)

Specifies a coordinate for the query defined by a longitude, and latitude.

The first position is the X coordinate, or longitude.

The second position is the Y coordinate, or latitude.

For example, position=xLongitude&position=yLatitude .

newSearchPlaceIndexForPosition

Arguments

:: Text
$sel:indexName:SearchPlaceIndexForPosition'

-> NonEmpty Double
$sel:position:SearchPlaceIndexForPosition'

-> SearchPlaceIndexForPosition

Create a value of SearchPlaceIndexForPosition with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:maxResults:SearchPlaceIndexForPosition', searchPlaceIndexForPosition_maxResults - An optional paramer. The maximum number of results returned per request.

Default value: 50

$sel:indexName:SearchPlaceIndexForPosition', searchPlaceIndexForPosition_indexName - The name of the place index resource you want to use for the search.

$sel:position:SearchPlaceIndexForPosition', searchPlaceIndexForPosition_position - Specifies a coordinate for the query defined by a longitude, and latitude.

The first position is the X coordinate, or longitude.

The second position is the Y coordinate, or latitude.

For example, position=xLongitude&position=yLatitude .

Request Lenses

searchPlaceIndexForPosition_maxResults :: Lens' SearchPlaceIndexForPosition (Maybe Natural)

An optional paramer. The maximum number of results returned per request.

Default value: 50

searchPlaceIndexForPosition_indexName :: Lens' SearchPlaceIndexForPosition Text

The name of the place index resource you want to use for the search.

searchPlaceIndexForPosition_position :: Lens' SearchPlaceIndexForPosition (NonEmpty Double)

Specifies a coordinate for the query defined by a longitude, and latitude.

The first position is the X coordinate, or longitude.

The second position is the Y coordinate, or latitude.

For example, position=xLongitude&position=yLatitude .

Destructuring the Response

data SearchPlaceIndexForPositionResponse

See: newSearchPlaceIndexForPositionResponse smart constructor.

Constructors

SearchPlaceIndexForPositionResponse'

Fields

httpStatus :: Int

The response's http status code.

results :: [SearchForPositionResult]

Returns a list of Places closest to the specified position. Each result contains additional information about the Places returned.

summary :: SearchPlaceIndexForPositionSummary

Contains a summary of the request.

newSearchPlaceIndexForPositionResponse

Arguments

:: Int
$sel:httpStatus:SearchPlaceIndexForPositionResponse'

-> SearchPlaceIndexForPositionSummary
$sel:summary:SearchPlaceIndexForPositionResponse'

-> SearchPlaceIndexForPositionResponse

Create a value of SearchPlaceIndexForPositionResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:httpStatus:SearchPlaceIndexForPositionResponse', searchPlaceIndexForPositionResponse_httpStatus - The response's http status code.

$sel:results:SearchPlaceIndexForPositionResponse', searchPlaceIndexForPositionResponse_results - Returns a list of Places closest to the specified position. Each result contains additional information about the Places returned.

$sel:summary:SearchPlaceIndexForPositionResponse', searchPlaceIndexForPositionResponse_summary - Contains a summary of the request.

Response Lenses

searchPlaceIndexForPositionResponse_httpStatus :: Lens' SearchPlaceIndexForPositionResponse Int

The response's http status code.

searchPlaceIndexForPositionResponse_results :: Lens' SearchPlaceIndexForPositionResponse [SearchForPositionResult]

Returns a list of Places closest to the specified position. Each result contains additional information about the Places returned.

searchPlaceIndexForPositionResponse_summary :: Lens' SearchPlaceIndexForPositionResponse SearchPlaceIndexForPositionSummary

Contains a summary of the request.