Amazonka.Location.Types.BatchDeleteGeofenceError

Description

Documentation

data BatchDeleteGeofenceError

Contains error details for each geofence that failed to delete from the geofence collection.

See: newBatchDeleteGeofenceError smart constructor.

Constructors

BatchDeleteGeofenceError'

Fields

error :: BatchItemError

Contains details associated to the batch error.

geofenceId :: Text

The geofence associated with the error message.

newBatchDeleteGeofenceError

Arguments

:: BatchItemError
$sel:error:BatchDeleteGeofenceError'

-> Text
$sel:geofenceId:BatchDeleteGeofenceError'

-> BatchDeleteGeofenceError

Create a value of BatchDeleteGeofenceError with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:error:BatchDeleteGeofenceError', batchDeleteGeofenceError_error - Contains details associated to the batch error.

$sel:geofenceId:BatchDeleteGeofenceError', batchDeleteGeofenceError_geofenceId - The geofence associated with the error message.

batchDeleteGeofenceError_error :: Lens' BatchDeleteGeofenceError BatchItemError

Contains details associated to the batch error.

batchDeleteGeofenceError_geofenceId :: Lens' BatchDeleteGeofenceError Text

The geofence associated with the error message.