Amazonka.Location.Types.DataSourceConfiguration

Description

Documentation

data DataSourceConfiguration

Specifies the data storage option chosen for requesting Places.

When using Amazon Location Places:

If using HERE Technologies as a data provider, you can't store results for locations in Japan by setting IntendedUse to Storage. parameter.

Under the MobileAssetTracking or MobilAssetManagement pricing plan, you can't store results from your place index resources by setting IntendedUse to Storage. This returns a validation exception error.

For more information, see the AWS Service Terms for Amazon Location Service.

See: newDataSourceConfiguration smart constructor.

Constructors

DataSourceConfiguration'

Fields

intendedUse :: Maybe IntendedUse

Specifies how the results of an operation will be stored by the caller.

Valid values include:

SingleUse specifies that the results won't be stored.

Storage specifies that the result can be cached or stored in a database.

Default value: SingleUse

newDataSourceConfiguration :: DataSourceConfiguration

Create a value of DataSourceConfiguration with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:intendedUse:DataSourceConfiguration', dataSourceConfiguration_intendedUse - Specifies how the results of an operation will be stored by the caller.

Valid values include:

SingleUse specifies that the results won't be stored.

Storage specifies that the result can be cached or stored in a database.

Default value: SingleUse

dataSourceConfiguration_intendedUse :: Lens' DataSourceConfiguration (Maybe IntendedUse)

Specifies how the results of an operation will be stored by the caller.

Valid values include:

SingleUse specifies that the results won't be stored.

Storage specifies that the result can be cached or stored in a database.

Default value: SingleUse