Amazonka.Location.Types.SearchPlaceIndexForTextSummary

Description

Documentation

data SearchPlaceIndexForTextSummary

A summary of the geocoding request sent using SearchPlaceIndexForText.

See: newSearchPlaceIndexForTextSummary smart constructor.

Constructors

SearchPlaceIndexForTextSummary'

Fields

filterBBox :: Maybe (Sensitive (NonEmpty Double))

Contains the coordinates for the optional bounding box coordinated entered in the geocoding request.

resultBBox :: Maybe (Sensitive (NonEmpty Double))

A bounding box that contains the search results within the specified area indicated by FilterBBox. A subset of bounding box specified using FilterBBox.

biasPosition :: Maybe (Sensitive (NonEmpty Double))

Contains the coordinates for the bias position entered in the geocoding request.

filterCountries :: Maybe (NonEmpty Text)

Contains the country filter entered in the geocoding request.

maxResults :: Maybe Natural

Contains the maximum number of results indicated for the request.

dataSource :: Text

The data provider of geospatial data. Indicates one of the available providers:

Esri

HERE

For additional details on data providers, see Amazon Location Service data providers.

text :: Sensitive Text

The address, name, city or region to be used in the geocoding request. In free-form text format. For example, Vancouver.

newSearchPlaceIndexForTextSummary

Arguments

:: Text
$sel:dataSource:SearchPlaceIndexForTextSummary'

-> Text
$sel:text:SearchPlaceIndexForTextSummary'

-> SearchPlaceIndexForTextSummary

Create a value of SearchPlaceIndexForTextSummary with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:filterBBox:SearchPlaceIndexForTextSummary', searchPlaceIndexForTextSummary_filterBBox - Contains the coordinates for the optional bounding box coordinated entered in the geocoding request.

$sel:resultBBox:SearchPlaceIndexForTextSummary', searchPlaceIndexForTextSummary_resultBBox - A bounding box that contains the search results within the specified area indicated by FilterBBox. A subset of bounding box specified using FilterBBox.

$sel:biasPosition:SearchPlaceIndexForTextSummary', searchPlaceIndexForTextSummary_biasPosition - Contains the coordinates for the bias position entered in the geocoding request.

$sel:filterCountries:SearchPlaceIndexForTextSummary', searchPlaceIndexForTextSummary_filterCountries - Contains the country filter entered in the geocoding request.

$sel:maxResults:SearchPlaceIndexForTextSummary', searchPlaceIndexForTextSummary_maxResults - Contains the maximum number of results indicated for the request.

$sel:dataSource:SearchPlaceIndexForTextSummary', searchPlaceIndexForTextSummary_dataSource - The data provider of geospatial data. Indicates one of the available providers:

Esri

HERE

For additional details on data providers, see Amazon Location Service data providers.

$sel:text:SearchPlaceIndexForTextSummary', searchPlaceIndexForTextSummary_text - The address, name, city or region to be used in the geocoding request. In free-form text format. For example, Vancouver.

searchPlaceIndexForTextSummary_filterBBox :: Lens' SearchPlaceIndexForTextSummary (Maybe (NonEmpty Double))

Contains the coordinates for the optional bounding box coordinated entered in the geocoding request.

searchPlaceIndexForTextSummary_resultBBox :: Lens' SearchPlaceIndexForTextSummary (Maybe (NonEmpty Double))

A bounding box that contains the search results within the specified area indicated by FilterBBox. A subset of bounding box specified using FilterBBox.

searchPlaceIndexForTextSummary_biasPosition :: Lens' SearchPlaceIndexForTextSummary (Maybe (NonEmpty Double))

Contains the coordinates for the bias position entered in the geocoding request.

searchPlaceIndexForTextSummary_filterCountries :: Lens' SearchPlaceIndexForTextSummary (Maybe (NonEmpty Text))

Contains the country filter entered in the geocoding request.

searchPlaceIndexForTextSummary_maxResults :: Lens' SearchPlaceIndexForTextSummary (Maybe Natural)

Contains the maximum number of results indicated for the request.

searchPlaceIndexForTextSummary_dataSource :: Lens' SearchPlaceIndexForTextSummary Text

The data provider of geospatial data. Indicates one of the available providers:

Esri

HERE

For additional details on data providers, see Amazon Location Service data providers.

searchPlaceIndexForTextSummary_text :: Lens' SearchPlaceIndexForTextSummary Text

The address, name, city or region to be used in the geocoding request. In free-form text format. For example, Vancouver.