Amazonka.Location.UpdateMap

Description

Updates the specified properties of a given map resource.

Creating a Request

data UpdateMap

See: newUpdateMap smart constructor.

Constructors

UpdateMap'

Fields

pricingPlan :: Maybe PricingPlan

Updates the pricing plan for the map resource.

For more information about each pricing plan option restrictions, see Amazon Location Service pricing.

description :: Maybe Text

Updates the description for the map resource.

mapName :: Text

The name of the map resource to update.

newUpdateMap

Arguments

:: Text
$sel:mapName:UpdateMap'

-> UpdateMap

Create a value of UpdateMap with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:pricingPlan:UpdateMap', updateMap_pricingPlan - Updates the pricing plan for the map resource.

For more information about each pricing plan option restrictions, see Amazon Location Service pricing.

$sel:description:UpdateMap', updateMap_description - Updates the description for the map resource.

$sel:mapName:UpdateMap', updateMap_mapName - The name of the map resource to update.

Request Lenses

updateMap_pricingPlan :: Lens' UpdateMap (Maybe PricingPlan)

Updates the pricing plan for the map resource.

For more information about each pricing plan option restrictions, see Amazon Location Service pricing.

updateMap_description :: Lens' UpdateMap (Maybe Text)

Updates the description for the map resource.

updateMap_mapName :: Lens' UpdateMap Text

The name of the map resource to update.

Destructuring the Response

data UpdateMapResponse

See: newUpdateMapResponse smart constructor.

Constructors

UpdateMapResponse'

Fields

httpStatus :: Int

The response's http status code.

mapArn :: Text

The Amazon Resource Name (ARN) of the updated map resource. Used to specify a resource across AWS.

Format example: arn:aws:geo:region:account-id:maps/ExampleMap

mapName :: Text

The name of the updated map resource.

updateTime :: POSIX

The timestamp for when the map resource was last updated in ISO 8601 format: YYYY-MM-DDThh:mm:ss.sssZ.

newUpdateMapResponse

Arguments

:: Int
$sel:httpStatus:UpdateMapResponse'

-> Text
$sel:mapArn:UpdateMapResponse'

-> Text
$sel:mapName:UpdateMap'

-> UTCTime
$sel:updateTime:UpdateMapResponse'

-> UpdateMapResponse

Create a value of UpdateMapResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:httpStatus:UpdateMapResponse', updateMapResponse_httpStatus - The response's http status code.

$sel:mapArn:UpdateMapResponse', updateMapResponse_mapArn - The Amazon Resource Name (ARN) of the updated map resource. Used to specify a resource across AWS.

Format example: arn:aws:geo:region:account-id:maps/ExampleMap

$sel:mapName:UpdateMap', updateMapResponse_mapName - The name of the updated map resource.

$sel:updateTime:UpdateMapResponse', updateMapResponse_updateTime - The timestamp for when the map resource was last updated in ISO 8601 format: YYYY-MM-DDThh:mm:ss.sssZ.

Response Lenses

updateMapResponse_httpStatus :: Lens' UpdateMapResponse Int

The response's http status code.

updateMapResponse_mapArn :: Lens' UpdateMapResponse Text

The Amazon Resource Name (ARN) of the updated map resource. Used to specify a resource across AWS.

Format example: arn:aws:geo:region:account-id:maps/ExampleMap

updateMapResponse_mapName :: Lens' UpdateMapResponse Text

The name of the updated map resource.

updateMapResponse_updateTime :: Lens' UpdateMapResponse UTCTime

The timestamp for when the map resource was last updated in ISO 8601 format: YYYY-MM-DDThh:mm:ss.sssZ.