Amazonka.Location.UpdateTracker

Description

Updates the specified properties of a given tracker resource.

Creating a Request

data UpdateTracker

See: newUpdateTracker smart constructor.

Constructors

UpdateTracker'

Fields

pricingPlan :: Maybe PricingPlan

Updates the pricing plan for the tracker resource.

For more information about each pricing plan option restrictions, see Amazon Location Service pricing.

pricingPlanDataSource :: Maybe Text

Updates the data provider for the tracker resource.

A required value for the following pricing plans: MobileAssetTracking| MobileAssetManagement

For more information about data providers and pricing plans, see the Amazon Location Service product page

This can only be updated when updating the PricingPlan in the same request.

Amazon Location Service uses PricingPlanDataSource to calculate billing for your tracker resource. Your data won't be shared with the data provider, and will remain in your AWS account and Region unless you move it.

description :: Maybe Text

Updates the description for the tracker resource.

positionFiltering :: Maybe PositionFiltering

Updates the position filtering for the tracker resource.

Valid values:

TimeBased - Location updates are evaluated against linked geofence collections, but not every location update is stored. If your update frequency is more often than 30 seconds, only one update per 30 seconds is stored for each unique device ID.

DistanceBased - If the device has moved less than 30 m (98.4 ft), location updates are ignored. Location updates within this distance are neither evaluated against linked geofence collections, nor stored. This helps control costs by reducing the number of geofence evaluations and device positions to retrieve. Distance-based filtering can also reduce the jitter effect when displaying device trajectory on a map.

trackerName :: Text

The name of the tracker resource to update.

newUpdateTracker

Arguments

:: Text
$sel:trackerName:UpdateTracker'

-> UpdateTracker

Create a value of UpdateTracker with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:pricingPlan:UpdateTracker', updateTracker_pricingPlan - Updates the pricing plan for the tracker resource.

For more information about each pricing plan option restrictions, see Amazon Location Service pricing.

$sel:pricingPlanDataSource:UpdateTracker', updateTracker_pricingPlanDataSource - Updates the data provider for the tracker resource.

A required value for the following pricing plans: MobileAssetTracking| MobileAssetManagement

For more information about data providers and pricing plans, see the Amazon Location Service product page

This can only be updated when updating the PricingPlan in the same request.

Amazon Location Service uses PricingPlanDataSource to calculate billing for your tracker resource. Your data won't be shared with the data provider, and will remain in your AWS account and Region unless you move it.

$sel:description:UpdateTracker', updateTracker_description - Updates the description for the tracker resource.

$sel:positionFiltering:UpdateTracker', updateTracker_positionFiltering - Updates the position filtering for the tracker resource.

Valid values:

TimeBased - Location updates are evaluated against linked geofence collections, but not every location update is stored. If your update frequency is more often than 30 seconds, only one update per 30 seconds is stored for each unique device ID.

DistanceBased - If the device has moved less than 30 m (98.4 ft), location updates are ignored. Location updates within this distance are neither evaluated against linked geofence collections, nor stored. This helps control costs by reducing the number of geofence evaluations and device positions to retrieve. Distance-based filtering can also reduce the jitter effect when displaying device trajectory on a map.

$sel:trackerName:UpdateTracker', updateTracker_trackerName - The name of the tracker resource to update.

Request Lenses

updateTracker_pricingPlan :: Lens' UpdateTracker (Maybe PricingPlan)

Updates the pricing plan for the tracker resource.

For more information about each pricing plan option restrictions, see Amazon Location Service pricing.

updateTracker_pricingPlanDataSource :: Lens' UpdateTracker (Maybe Text)

Updates the data provider for the tracker resource.

A required value for the following pricing plans: MobileAssetTracking| MobileAssetManagement

For more information about data providers and pricing plans, see the Amazon Location Service product page

This can only be updated when updating the PricingPlan in the same request.

Amazon Location Service uses PricingPlanDataSource to calculate billing for your tracker resource. Your data won't be shared with the data provider, and will remain in your AWS account and Region unless you move it.

updateTracker_description :: Lens' UpdateTracker (Maybe Text)

Updates the description for the tracker resource.

updateTracker_positionFiltering :: Lens' UpdateTracker (Maybe PositionFiltering)

Updates the position filtering for the tracker resource.

Valid values:

TimeBased - Location updates are evaluated against linked geofence collections, but not every location update is stored. If your update frequency is more often than 30 seconds, only one update per 30 seconds is stored for each unique device ID.

DistanceBased - If the device has moved less than 30 m (98.4 ft), location updates are ignored. Location updates within this distance are neither evaluated against linked geofence collections, nor stored. This helps control costs by reducing the number of geofence evaluations and device positions to retrieve. Distance-based filtering can also reduce the jitter effect when displaying device trajectory on a map.

updateTracker_trackerName :: Lens' UpdateTracker Text

The name of the tracker resource to update.

Destructuring the Response

data UpdateTrackerResponse

See: newUpdateTrackerResponse smart constructor.

Constructors

UpdateTrackerResponse'

Fields

httpStatus :: Int

The response's http status code.

trackerArn :: Text

The Amazon Resource Name (ARN) of the updated tracker resource. Used to specify a resource across AWS.

Format example: arn:aws:geo:region:account-id:tracker/ExampleTracker

trackerName :: Text

The name of the updated tracker resource.

updateTime :: POSIX

The timestamp for when the tracker resource was last updated in ISO 8601 format: YYYY-MM-DDThh:mm:ss.sssZ.

newUpdateTrackerResponse

Arguments

:: Int
$sel:httpStatus:UpdateTrackerResponse'

-> Text
$sel:trackerArn:UpdateTrackerResponse'

-> Text
$sel:trackerName:UpdateTracker'

-> UTCTime
$sel:updateTime:UpdateTrackerResponse'

-> UpdateTrackerResponse

Create a value of UpdateTrackerResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:httpStatus:UpdateTrackerResponse', updateTrackerResponse_httpStatus - The response's http status code.

$sel:trackerArn:UpdateTrackerResponse', updateTrackerResponse_trackerArn - The Amazon Resource Name (ARN) of the updated tracker resource. Used to specify a resource across AWS.

Format example: arn:aws:geo:region:account-id:tracker/ExampleTracker

$sel:trackerName:UpdateTracker', updateTrackerResponse_trackerName - The name of the updated tracker resource.

$sel:updateTime:UpdateTrackerResponse', updateTrackerResponse_updateTime - The timestamp for when the tracker resource was last updated in ISO 8601 format: YYYY-MM-DDThh:mm:ss.sssZ.

Response Lenses

updateTrackerResponse_httpStatus :: Lens' UpdateTrackerResponse Int

The response's http status code.

updateTrackerResponse_trackerArn :: Lens' UpdateTrackerResponse Text

The Amazon Resource Name (ARN) of the updated tracker resource. Used to specify a resource across AWS.

Format example: arn:aws:geo:region:account-id:tracker/ExampleTracker

updateTrackerResponse_trackerName :: Lens' UpdateTrackerResponse Text

The name of the updated tracker resource.

updateTrackerResponse_updateTime :: Lens' UpdateTrackerResponse UTCTime

The timestamp for when the tracker resource was last updated in ISO 8601 format: YYYY-MM-DDThh:mm:ss.sssZ.