Amazonka.LookoutEquipment.CreateModel

Description

Creates an ML model for data inference.

A machine-learning (ML) model is a mathematical model that finds patterns in your data. In Amazon Lookout for Equipment, the model learns the patterns of normal behavior and detects abnormal behavior that could be potential equipment failure (or maintenance events). The models are made by analyzing normal data and abnormalities in machine behavior that have already occurred.

Your model is trained using a portion of the data from your dataset and uses that data to learn patterns of normal behavior and abnormal patterns that lead to equipment failure. Another portion of the data is used to evaluate the model's accuracy.

Creating a Request

data CreateModel

See: newCreateModel smart constructor.

Constructors

CreateModel'

Fields

dataPreProcessingConfiguration :: Maybe DataPreProcessingConfiguration

The configuration is the TargetSamplingRate, which is the sampling rate of the data after post processing by Amazon Lookout for Equipment. For example, if you provide data that has been collected at a 1 second level and you want the system to resample the data at a 1 minute rate before training, the TargetSamplingRate is 1 minute.

When providing a value for the TargetSamplingRate, you must attach the prefix "PT" to the rate you want. The value for a 1 second rate is therefore PT1S, the value for a 15 minute rate is PT15M, and the value for a 1 hour rate is PT1H

trainingDataEndTime :: Maybe POSIX

Indicates the time reference in the dataset that should be used to end the subset of training data for the ML model.

datasetSchema :: Maybe DatasetSchema

The data schema for the ML model being created.

evaluationDataStartTime :: Maybe POSIX

Indicates the time reference in the dataset that should be used to begin the subset of evaluation data for the ML model.

offCondition :: Maybe Text

Indicates that the asset associated with this sensor has been shut off. As long as this condition is met, Lookout for Equipment will not use data from this asset for training, evaluation, or inference.

evaluationDataEndTime :: Maybe POSIX

Indicates the time reference in the dataset that should be used to end the subset of evaluation data for the ML model.

trainingDataStartTime :: Maybe POSIX

Indicates the time reference in the dataset that should be used to begin the subset of training data for the ML model.

labelsInputConfiguration :: Maybe LabelsInputConfiguration

The input configuration for the labels being used for the ML model that's being created.

tags :: Maybe [Tag]

Any tags associated with the ML model being created.

serverSideKmsKeyId :: Maybe Text

Provides the identifier of the KMS key used to encrypt model data by Amazon Lookout for Equipment.

roleArn :: Maybe Text

The Amazon Resource Name (ARN) of a role with permission to access the data source being used to create the ML model.

modelName :: Text

The name for the ML model to be created.

datasetName :: Text

The name of the dataset for the ML model being created.

clientToken :: Text

A unique identifier for the request. If you do not set the client request token, Amazon Lookout for Equipment generates one.

newCreateModel

Arguments

:: Text
$sel:modelName:CreateModel'

-> Text
$sel:datasetName:CreateModel'

-> Text
$sel:clientToken:CreateModel'

-> CreateModel

Create a value of CreateModel with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:dataPreProcessingConfiguration:CreateModel', createModel_dataPreProcessingConfiguration - The configuration is the TargetSamplingRate, which is the sampling rate of the data after post processing by Amazon Lookout for Equipment. For example, if you provide data that has been collected at a 1 second level and you want the system to resample the data at a 1 minute rate before training, the TargetSamplingRate is 1 minute.

When providing a value for the TargetSamplingRate, you must attach the prefix "PT" to the rate you want. The value for a 1 second rate is therefore PT1S, the value for a 15 minute rate is PT15M, and the value for a 1 hour rate is PT1H

$sel:trainingDataEndTime:CreateModel', createModel_trainingDataEndTime - Indicates the time reference in the dataset that should be used to end the subset of training data for the ML model.

$sel:datasetSchema:CreateModel', createModel_datasetSchema - The data schema for the ML model being created.

$sel:evaluationDataStartTime:CreateModel', createModel_evaluationDataStartTime - Indicates the time reference in the dataset that should be used to begin the subset of evaluation data for the ML model.

$sel:offCondition:CreateModel', createModel_offCondition - Indicates that the asset associated with this sensor has been shut off. As long as this condition is met, Lookout for Equipment will not use data from this asset for training, evaluation, or inference.

$sel:evaluationDataEndTime:CreateModel', createModel_evaluationDataEndTime - Indicates the time reference in the dataset that should be used to end the subset of evaluation data for the ML model.

$sel:trainingDataStartTime:CreateModel', createModel_trainingDataStartTime - Indicates the time reference in the dataset that should be used to begin the subset of training data for the ML model.

$sel:labelsInputConfiguration:CreateModel', createModel_labelsInputConfiguration - The input configuration for the labels being used for the ML model that's being created.

$sel:tags:CreateModel', createModel_tags - Any tags associated with the ML model being created.

$sel:serverSideKmsKeyId:CreateModel', createModel_serverSideKmsKeyId - Provides the identifier of the KMS key used to encrypt model data by Amazon Lookout for Equipment.

$sel:roleArn:CreateModel', createModel_roleArn - The Amazon Resource Name (ARN) of a role with permission to access the data source being used to create the ML model.

$sel:modelName:CreateModel', createModel_modelName - The name for the ML model to be created.

$sel:datasetName:CreateModel', createModel_datasetName - The name of the dataset for the ML model being created.

$sel:clientToken:CreateModel', createModel_clientToken - A unique identifier for the request. If you do not set the client request token, Amazon Lookout for Equipment generates one.

Request Lenses

createModel_dataPreProcessingConfiguration :: Lens' CreateModel (Maybe DataPreProcessingConfiguration)

The configuration is the TargetSamplingRate, which is the sampling rate of the data after post processing by Amazon Lookout for Equipment. For example, if you provide data that has been collected at a 1 second level and you want the system to resample the data at a 1 minute rate before training, the TargetSamplingRate is 1 minute.

When providing a value for the TargetSamplingRate, you must attach the prefix "PT" to the rate you want. The value for a 1 second rate is therefore PT1S, the value for a 15 minute rate is PT15M, and the value for a 1 hour rate is PT1H

createModel_trainingDataEndTime :: Lens' CreateModel (Maybe UTCTime)

Indicates the time reference in the dataset that should be used to end the subset of training data for the ML model.

createModel_datasetSchema :: Lens' CreateModel (Maybe DatasetSchema)

The data schema for the ML model being created.

createModel_evaluationDataStartTime :: Lens' CreateModel (Maybe UTCTime)

Indicates the time reference in the dataset that should be used to begin the subset of evaluation data for the ML model.

createModel_offCondition :: Lens' CreateModel (Maybe Text)

Indicates that the asset associated with this sensor has been shut off. As long as this condition is met, Lookout for Equipment will not use data from this asset for training, evaluation, or inference.

createModel_evaluationDataEndTime :: Lens' CreateModel (Maybe UTCTime)

Indicates the time reference in the dataset that should be used to end the subset of evaluation data for the ML model.

createModel_trainingDataStartTime :: Lens' CreateModel (Maybe UTCTime)

Indicates the time reference in the dataset that should be used to begin the subset of training data for the ML model.

createModel_labelsInputConfiguration :: Lens' CreateModel (Maybe LabelsInputConfiguration)

The input configuration for the labels being used for the ML model that's being created.

createModel_tags :: Lens' CreateModel (Maybe [Tag])

Any tags associated with the ML model being created.

createModel_serverSideKmsKeyId :: Lens' CreateModel (Maybe Text)

Provides the identifier of the KMS key used to encrypt model data by Amazon Lookout for Equipment.

createModel_roleArn :: Lens' CreateModel (Maybe Text)

The Amazon Resource Name (ARN) of a role with permission to access the data source being used to create the ML model.

createModel_modelName :: Lens' CreateModel Text

The name for the ML model to be created.

createModel_datasetName :: Lens' CreateModel Text

The name of the dataset for the ML model being created.

createModel_clientToken :: Lens' CreateModel Text

A unique identifier for the request. If you do not set the client request token, Amazon Lookout for Equipment generates one.

Destructuring the Response

data CreateModelResponse

See: newCreateModelResponse smart constructor.

Constructors

CreateModelResponse'

Fields

status :: Maybe ModelStatus

Indicates the status of the CreateModel operation.

modelArn :: Maybe Text

The Amazon Resource Name (ARN) of the model being created.

httpStatus :: Int

The response's http status code.

newCreateModelResponse

Arguments

:: Int
$sel:httpStatus:CreateModelResponse'

-> CreateModelResponse

Create a value of CreateModelResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:status:CreateModelResponse', createModelResponse_status - Indicates the status of the CreateModel operation.

$sel:modelArn:CreateModelResponse', createModelResponse_modelArn - The Amazon Resource Name (ARN) of the model being created.

$sel:httpStatus:CreateModelResponse', createModelResponse_httpStatus - The response's http status code.

Response Lenses

createModelResponse_status :: Lens' CreateModelResponse (Maybe ModelStatus)

Indicates the status of the CreateModel operation.

createModelResponse_modelArn :: Lens' CreateModelResponse (Maybe Text)

The Amazon Resource Name (ARN) of the model being created.

createModelResponse_httpStatus :: Lens' CreateModelResponse Int

The response's http status code.