Amazonka.LookoutEquipment.DescribeInferenceScheduler

Description

Specifies information about the inference scheduler being used, including name, model, status, and associated metadata

Creating a Request

data DescribeInferenceScheduler

See: newDescribeInferenceScheduler smart constructor.

Constructors

DescribeInferenceScheduler'

Fields

inferenceSchedulerName :: Text

The name of the inference scheduler being described.

newDescribeInferenceScheduler

Arguments

:: Text
$sel:inferenceSchedulerName:DescribeInferenceScheduler'

-> DescribeInferenceScheduler

Create a value of DescribeInferenceScheduler with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:inferenceSchedulerName:DescribeInferenceScheduler', describeInferenceScheduler_inferenceSchedulerName - The name of the inference scheduler being described.

Request Lenses

describeInferenceScheduler_inferenceSchedulerName :: Lens' DescribeInferenceScheduler Text

The name of the inference scheduler being described.

Destructuring the Response

data DescribeInferenceSchedulerResponse

See: newDescribeInferenceSchedulerResponse smart constructor.

Constructors

DescribeInferenceSchedulerResponse'

Fields

status :: Maybe InferenceSchedulerStatus

Indicates the status of the inference scheduler.

dataUploadFrequency :: Maybe DataUploadFrequency

Specifies how often data is uploaded to the source S3 bucket for the input data. This value is the length of time between data uploads. For instance, if you select 5 minutes, Amazon Lookout for Equipment will upload the real-time data to the source bucket once every 5 minutes. This frequency also determines how often Amazon Lookout for Equipment starts a scheduled inference on your data. In this example, it starts once every 5 minutes.

dataDelayOffsetInMinutes :: Maybe Natural

A period of time (in minutes) by which inference on the data is delayed after the data starts. For instance, if you select an offset delay time of five minutes, inference will not begin on the data until the first data measurement after the five minute mark. For example, if five minutes is selected, the inference scheduler will wake up at the configured frequency with the additional five minute delay time to check the customer S3 bucket. The customer can upload data at the same frequency and they don't need to stop and restart the scheduler when uploading new data.

modelArn :: Maybe Text

The Amazon Resource Name (ARN) of the ML model of the inference scheduler being described.

createdAt :: Maybe POSIX

Specifies the time at which the inference scheduler was created.

modelName :: Maybe Text

The name of the ML model of the inference scheduler being described.

inferenceSchedulerArn :: Maybe Text

The Amazon Resource Name (ARN) of the inference scheduler being described.

dataOutputConfiguration :: Maybe InferenceOutputConfiguration

Specifies information for the output results for the inference scheduler, including the output S3 location.

updatedAt :: Maybe POSIX

Specifies the time at which the inference scheduler was last updated, if it was.

inferenceSchedulerName :: Maybe Text

The name of the inference scheduler being described.

dataInputConfiguration :: Maybe InferenceInputConfiguration

Specifies configuration information for the input data for the inference scheduler, including delimiter, format, and dataset location.

serverSideKmsKeyId :: Maybe Text

Provides the identifier of the KMS key used to encrypt inference scheduler data by Amazon Lookout for Equipment.

roleArn :: Maybe Text

The Amazon Resource Name (ARN) of a role with permission to access the data source for the inference scheduler being described.

httpStatus :: Int

The response's http status code.

newDescribeInferenceSchedulerResponse

Arguments

:: Int
$sel:httpStatus:DescribeInferenceSchedulerResponse'

-> DescribeInferenceSchedulerResponse

Create a value of DescribeInferenceSchedulerResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:status:DescribeInferenceSchedulerResponse', describeInferenceSchedulerResponse_status - Indicates the status of the inference scheduler.

$sel:dataUploadFrequency:DescribeInferenceSchedulerResponse', describeInferenceSchedulerResponse_dataUploadFrequency - Specifies how often data is uploaded to the source S3 bucket for the input data. This value is the length of time between data uploads. For instance, if you select 5 minutes, Amazon Lookout for Equipment will upload the real-time data to the source bucket once every 5 minutes. This frequency also determines how often Amazon Lookout for Equipment starts a scheduled inference on your data. In this example, it starts once every 5 minutes.

$sel:dataDelayOffsetInMinutes:DescribeInferenceSchedulerResponse', describeInferenceSchedulerResponse_dataDelayOffsetInMinutes - A period of time (in minutes) by which inference on the data is delayed after the data starts. For instance, if you select an offset delay time of five minutes, inference will not begin on the data until the first data measurement after the five minute mark. For example, if five minutes is selected, the inference scheduler will wake up at the configured frequency with the additional five minute delay time to check the customer S3 bucket. The customer can upload data at the same frequency and they don't need to stop and restart the scheduler when uploading new data.

$sel:modelArn:DescribeInferenceSchedulerResponse', describeInferenceSchedulerResponse_modelArn - The Amazon Resource Name (ARN) of the ML model of the inference scheduler being described.

$sel:createdAt:DescribeInferenceSchedulerResponse', describeInferenceSchedulerResponse_createdAt - Specifies the time at which the inference scheduler was created.

$sel:modelName:DescribeInferenceSchedulerResponse', describeInferenceSchedulerResponse_modelName - The name of the ML model of the inference scheduler being described.

$sel:inferenceSchedulerArn:DescribeInferenceSchedulerResponse', describeInferenceSchedulerResponse_inferenceSchedulerArn - The Amazon Resource Name (ARN) of the inference scheduler being described.

$sel:dataOutputConfiguration:DescribeInferenceSchedulerResponse', describeInferenceSchedulerResponse_dataOutputConfiguration - Specifies information for the output results for the inference scheduler, including the output S3 location.

$sel:updatedAt:DescribeInferenceSchedulerResponse', describeInferenceSchedulerResponse_updatedAt - Specifies the time at which the inference scheduler was last updated, if it was.

$sel:inferenceSchedulerName:DescribeInferenceScheduler', describeInferenceSchedulerResponse_inferenceSchedulerName - The name of the inference scheduler being described.

$sel:dataInputConfiguration:DescribeInferenceSchedulerResponse', describeInferenceSchedulerResponse_dataInputConfiguration - Specifies configuration information for the input data for the inference scheduler, including delimiter, format, and dataset location.

$sel:serverSideKmsKeyId:DescribeInferenceSchedulerResponse', describeInferenceSchedulerResponse_serverSideKmsKeyId - Provides the identifier of the KMS key used to encrypt inference scheduler data by Amazon Lookout for Equipment.

$sel:roleArn:DescribeInferenceSchedulerResponse', describeInferenceSchedulerResponse_roleArn - The Amazon Resource Name (ARN) of a role with permission to access the data source for the inference scheduler being described.

$sel:httpStatus:DescribeInferenceSchedulerResponse', describeInferenceSchedulerResponse_httpStatus - The response's http status code.

Response Lenses

describeInferenceSchedulerResponse_status :: Lens' DescribeInferenceSchedulerResponse (Maybe InferenceSchedulerStatus)

Indicates the status of the inference scheduler.

describeInferenceSchedulerResponse_dataUploadFrequency :: Lens' DescribeInferenceSchedulerResponse (Maybe DataUploadFrequency)

Specifies how often data is uploaded to the source S3 bucket for the input data. This value is the length of time between data uploads. For instance, if you select 5 minutes, Amazon Lookout for Equipment will upload the real-time data to the source bucket once every 5 minutes. This frequency also determines how often Amazon Lookout for Equipment starts a scheduled inference on your data. In this example, it starts once every 5 minutes.

describeInferenceSchedulerResponse_dataDelayOffsetInMinutes :: Lens' DescribeInferenceSchedulerResponse (Maybe Natural)

A period of time (in minutes) by which inference on the data is delayed after the data starts. For instance, if you select an offset delay time of five minutes, inference will not begin on the data until the first data measurement after the five minute mark. For example, if five minutes is selected, the inference scheduler will wake up at the configured frequency with the additional five minute delay time to check the customer S3 bucket. The customer can upload data at the same frequency and they don't need to stop and restart the scheduler when uploading new data.

describeInferenceSchedulerResponse_modelArn :: Lens' DescribeInferenceSchedulerResponse (Maybe Text)

The Amazon Resource Name (ARN) of the ML model of the inference scheduler being described.

describeInferenceSchedulerResponse_createdAt :: Lens' DescribeInferenceSchedulerResponse (Maybe UTCTime)

Specifies the time at which the inference scheduler was created.

describeInferenceSchedulerResponse_modelName :: Lens' DescribeInferenceSchedulerResponse (Maybe Text)

The name of the ML model of the inference scheduler being described.

describeInferenceSchedulerResponse_inferenceSchedulerArn :: Lens' DescribeInferenceSchedulerResponse (Maybe Text)

The Amazon Resource Name (ARN) of the inference scheduler being described.

describeInferenceSchedulerResponse_dataOutputConfiguration :: Lens' DescribeInferenceSchedulerResponse (Maybe InferenceOutputConfiguration)

Specifies information for the output results for the inference scheduler, including the output S3 location.

describeInferenceSchedulerResponse_updatedAt :: Lens' DescribeInferenceSchedulerResponse (Maybe UTCTime)

Specifies the time at which the inference scheduler was last updated, if it was.

describeInferenceSchedulerResponse_inferenceSchedulerName :: Lens' DescribeInferenceSchedulerResponse (Maybe Text)

The name of the inference scheduler being described.

describeInferenceSchedulerResponse_dataInputConfiguration :: Lens' DescribeInferenceSchedulerResponse (Maybe InferenceInputConfiguration)

Specifies configuration information for the input data for the inference scheduler, including delimiter, format, and dataset location.

describeInferenceSchedulerResponse_serverSideKmsKeyId :: Lens' DescribeInferenceSchedulerResponse (Maybe Text)

Provides the identifier of the KMS key used to encrypt inference scheduler data by Amazon Lookout for Equipment.

describeInferenceSchedulerResponse_roleArn :: Lens' DescribeInferenceSchedulerResponse (Maybe Text)

The Amazon Resource Name (ARN) of a role with permission to access the data source for the inference scheduler being described.

describeInferenceSchedulerResponse_httpStatus :: Lens' DescribeInferenceSchedulerResponse Int

The response's http status code.