Amazonka.LookoutEquipment.ListInferenceSchedulers

Description

Retrieves a list of all inference schedulers currently available for your account.

Creating a Request

data ListInferenceSchedulers

See: newListInferenceSchedulers smart constructor.

Constructors

ListInferenceSchedulers'

Fields

modelName :: Maybe Text

The name of the ML model used by the inference scheduler to be listed.

nextToken :: Maybe Text

An opaque pagination token indicating where to continue the listing of inference schedulers.

inferenceSchedulerNameBeginsWith :: Maybe Text

The beginning of the name of the inference schedulers to be listed.

maxResults :: Maybe Natural

Specifies the maximum number of inference schedulers to list.

newListInferenceSchedulers :: ListInferenceSchedulers

Create a value of ListInferenceSchedulers with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:modelName:ListInferenceSchedulers', listInferenceSchedulers_modelName - The name of the ML model used by the inference scheduler to be listed.

$sel:nextToken:ListInferenceSchedulers', listInferenceSchedulers_nextToken - An opaque pagination token indicating where to continue the listing of inference schedulers.

$sel:inferenceSchedulerNameBeginsWith:ListInferenceSchedulers', listInferenceSchedulers_inferenceSchedulerNameBeginsWith - The beginning of the name of the inference schedulers to be listed.

$sel:maxResults:ListInferenceSchedulers', listInferenceSchedulers_maxResults - Specifies the maximum number of inference schedulers to list.

Request Lenses

listInferenceSchedulers_modelName :: Lens' ListInferenceSchedulers (Maybe Text)

The name of the ML model used by the inference scheduler to be listed.

listInferenceSchedulers_nextToken :: Lens' ListInferenceSchedulers (Maybe Text)

An opaque pagination token indicating where to continue the listing of inference schedulers.

listInferenceSchedulers_inferenceSchedulerNameBeginsWith :: Lens' ListInferenceSchedulers (Maybe Text)

The beginning of the name of the inference schedulers to be listed.

listInferenceSchedulers_maxResults :: Lens' ListInferenceSchedulers (Maybe Natural)

Specifies the maximum number of inference schedulers to list.

Destructuring the Response

data ListInferenceSchedulersResponse

See: newListInferenceSchedulersResponse smart constructor.

Constructors

ListInferenceSchedulersResponse'

Fields

inferenceSchedulerSummaries :: Maybe [InferenceSchedulerSummary]

Provides information about the specified inference scheduler, including data upload frequency, model name and ARN, and status.

nextToken :: Maybe Text

An opaque pagination token indicating where to continue the listing of inference schedulers.

httpStatus :: Int

The response's http status code.

newListInferenceSchedulersResponse

Arguments

:: Int
$sel:httpStatus:ListInferenceSchedulersResponse'

-> ListInferenceSchedulersResponse

Create a value of ListInferenceSchedulersResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:inferenceSchedulerSummaries:ListInferenceSchedulersResponse', listInferenceSchedulersResponse_inferenceSchedulerSummaries - Provides information about the specified inference scheduler, including data upload frequency, model name and ARN, and status.

$sel:nextToken:ListInferenceSchedulers', listInferenceSchedulersResponse_nextToken - An opaque pagination token indicating where to continue the listing of inference schedulers.

$sel:httpStatus:ListInferenceSchedulersResponse', listInferenceSchedulersResponse_httpStatus - The response's http status code.

Response Lenses

listInferenceSchedulersResponse_inferenceSchedulerSummaries :: Lens' ListInferenceSchedulersResponse (Maybe [InferenceSchedulerSummary])

Provides information about the specified inference scheduler, including data upload frequency, model name and ARN, and status.

listInferenceSchedulersResponse_nextToken :: Lens' ListInferenceSchedulersResponse (Maybe Text)

An opaque pagination token indicating where to continue the listing of inference schedulers.

listInferenceSchedulersResponse_httpStatus :: Lens' ListInferenceSchedulersResponse Int

The response's http status code.