Amazonka.LookoutEquipment.Types

Description

Service Configuration

defaultService :: Service

API version 2020-12-15 of the Amazon Lookout for Equipment SDK configuration.

Errors

_ValidationException :: AsError a => Getting (First ServiceError) a ServiceError

The input fails to satisfy constraints specified by Amazon Lookout for Equipment or a related AWS service that's being utilized.

_AccessDeniedException :: AsError a => Getting (First ServiceError) a ServiceError

The request could not be completed because you do not have access to the resource.

_ConflictException :: AsError a => Getting (First ServiceError) a ServiceError

The request could not be completed due to a conflict with the current state of the target resource.

_ServiceQuotaExceededException :: AsError a => Getting (First ServiceError) a ServiceError

Resource limitations have been exceeded.

_ThrottlingException :: AsError a => Getting (First ServiceError) a ServiceError

The request was denied due to request throttling.

_InternalServerException :: AsError a => Getting (First ServiceError) a ServiceError

Processing of the request has failed because of an unknown error, exception or failure.

_ResourceNotFoundException :: AsError a => Getting (First ServiceError) a ServiceError

The resource requested could not be found. Verify the resource ID and retry your request.

DataUploadFrequency

newtype DataUploadFrequency

Constructors

DataUploadFrequency'

Fields

fromDataUploadFrequency :: Text

Bundled Patterns

pattern DataUploadFrequency_PT10M :: DataUploadFrequency

pattern DataUploadFrequency_PT15M :: DataUploadFrequency

pattern DataUploadFrequency_PT1H :: DataUploadFrequency

pattern DataUploadFrequency_PT30M :: DataUploadFrequency

pattern DataUploadFrequency_PT5M :: DataUploadFrequency

DatasetStatus

newtype DatasetStatus

Constructors

DatasetStatus'

Fields

fromDatasetStatus :: Text

Bundled Patterns

pattern DatasetStatus_ACTIVE :: DatasetStatus

pattern DatasetStatus_CREATED :: DatasetStatus

pattern DatasetStatus_INGESTION_IN_PROGRESS :: DatasetStatus

InferenceExecutionStatus

newtype InferenceExecutionStatus

Constructors

InferenceExecutionStatus'

Fields

fromInferenceExecutionStatus :: Text

Bundled Patterns

pattern InferenceExecutionStatus_FAILED :: InferenceExecutionStatus

pattern InferenceExecutionStatus_IN_PROGRESS :: InferenceExecutionStatus

pattern InferenceExecutionStatus_SUCCESS :: InferenceExecutionStatus

InferenceSchedulerStatus

newtype InferenceSchedulerStatus

Constructors

InferenceSchedulerStatus'

Fields

fromInferenceSchedulerStatus :: Text

Bundled Patterns

pattern InferenceSchedulerStatus_PENDING :: InferenceSchedulerStatus

pattern InferenceSchedulerStatus_RUNNING :: InferenceSchedulerStatus

pattern InferenceSchedulerStatus_STOPPED :: InferenceSchedulerStatus

pattern InferenceSchedulerStatus_STOPPING :: InferenceSchedulerStatus

IngestionJobStatus

newtype IngestionJobStatus

Constructors

IngestionJobStatus'

Fields

fromIngestionJobStatus :: Text

Bundled Patterns

pattern IngestionJobStatus_FAILED :: IngestionJobStatus

pattern IngestionJobStatus_IN_PROGRESS :: IngestionJobStatus

pattern IngestionJobStatus_SUCCESS :: IngestionJobStatus

ModelStatus

newtype ModelStatus

Constructors

ModelStatus'

Fields

fromModelStatus :: Text

Bundled Patterns

pattern ModelStatus_FAILED :: ModelStatus

pattern ModelStatus_IN_PROGRESS :: ModelStatus

pattern ModelStatus_SUCCESS :: ModelStatus

TargetSamplingRate

newtype TargetSamplingRate

Constructors

TargetSamplingRate'

Fields

fromTargetSamplingRate :: Text

Bundled Patterns

pattern TargetSamplingRate_PT10M :: TargetSamplingRate

pattern TargetSamplingRate_PT10S :: TargetSamplingRate

pattern TargetSamplingRate_PT15M :: TargetSamplingRate

pattern TargetSamplingRate_PT15S :: TargetSamplingRate

pattern TargetSamplingRate_PT1H :: TargetSamplingRate

pattern TargetSamplingRate_PT1M :: TargetSamplingRate

pattern TargetSamplingRate_PT1S :: TargetSamplingRate

pattern TargetSamplingRate_PT30M :: TargetSamplingRate

pattern TargetSamplingRate_PT30S :: TargetSamplingRate

pattern TargetSamplingRate_PT5M :: TargetSamplingRate

pattern TargetSamplingRate_PT5S :: TargetSamplingRate

DataIngestionJobSummary

data DataIngestionJobSummary

Provides information about a specified data ingestion job, including dataset information, data ingestion configuration, and status.

See: newDataIngestionJobSummary smart constructor.

Constructors

DataIngestionJobSummary'

Fields

ingestionInputConfiguration :: Maybe IngestionInputConfiguration

Specifies information for the input data for the data inference job, including data S3 location parameters.

status :: Maybe IngestionJobStatus

Indicates the status of the data ingestion job.

datasetArn :: Maybe Text

The Amazon Resource Name (ARN) of the dataset used in the data ingestion job.

jobId :: Maybe Text

Indicates the job ID of the data ingestion job.

datasetName :: Maybe Text

The name of the dataset used for the data ingestion job.

newDataIngestionJobSummary :: DataIngestionJobSummary

Create a value of DataIngestionJobSummary with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:ingestionInputConfiguration:DataIngestionJobSummary', dataIngestionJobSummary_ingestionInputConfiguration - Specifies information for the input data for the data inference job, including data S3 location parameters.

$sel:status:DataIngestionJobSummary', dataIngestionJobSummary_status - Indicates the status of the data ingestion job.

$sel:datasetArn:DataIngestionJobSummary', dataIngestionJobSummary_datasetArn - The Amazon Resource Name (ARN) of the dataset used in the data ingestion job.

$sel:jobId:DataIngestionJobSummary', dataIngestionJobSummary_jobId - Indicates the job ID of the data ingestion job.

$sel:datasetName:DataIngestionJobSummary', dataIngestionJobSummary_datasetName - The name of the dataset used for the data ingestion job.

dataIngestionJobSummary_ingestionInputConfiguration :: Lens' DataIngestionJobSummary (Maybe IngestionInputConfiguration)

Specifies information for the input data for the data inference job, including data S3 location parameters.

dataIngestionJobSummary_status :: Lens' DataIngestionJobSummary (Maybe IngestionJobStatus)

Indicates the status of the data ingestion job.

dataIngestionJobSummary_datasetArn :: Lens' DataIngestionJobSummary (Maybe Text)

The Amazon Resource Name (ARN) of the dataset used in the data ingestion job.

dataIngestionJobSummary_jobId :: Lens' DataIngestionJobSummary (Maybe Text)

Indicates the job ID of the data ingestion job.

dataIngestionJobSummary_datasetName :: Lens' DataIngestionJobSummary (Maybe Text)

The name of the dataset used for the data ingestion job.

DataPreProcessingConfiguration

data DataPreProcessingConfiguration

The configuration is the TargetSamplingRate, which is the sampling rate of the data after post processing by Amazon Lookout for Equipment. For example, if you provide data that has been collected at a 1 second level and you want the system to resample the data at a 1 minute rate before training, the TargetSamplingRate is 1 minute.

When providing a value for the TargetSamplingRate, you must attach the prefix "PT" to the rate you want. The value for a 1 second rate is therefore PT1S, the value for a 15 minute rate is PT15M, and the value for a 1 hour rate is PT1H

See: newDataPreProcessingConfiguration smart constructor.

Constructors

DataPreProcessingConfiguration'

Fields

targetSamplingRate :: Maybe TargetSamplingRate

The sampling rate of the data after post processing by Amazon Lookout for Equipment. For example, if you provide data that has been collected at a 1 second level and you want the system to resample the data at a 1 minute rate before training, the TargetSamplingRate is 1 minute.

When providing a value for the TargetSamplingRate, you must attach the prefix "PT" to the rate you want. The value for a 1 second rate is therefore PT1S, the value for a 15 minute rate is PT15M, and the value for a 1 hour rate is PT1H

newDataPreProcessingConfiguration :: DataPreProcessingConfiguration

Create a value of DataPreProcessingConfiguration with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:targetSamplingRate:DataPreProcessingConfiguration', dataPreProcessingConfiguration_targetSamplingRate - The sampling rate of the data after post processing by Amazon Lookout for Equipment. For example, if you provide data that has been collected at a 1 second level and you want the system to resample the data at a 1 minute rate before training, the TargetSamplingRate is 1 minute.

When providing a value for the TargetSamplingRate, you must attach the prefix "PT" to the rate you want. The value for a 1 second rate is therefore PT1S, the value for a 15 minute rate is PT15M, and the value for a 1 hour rate is PT1H

dataPreProcessingConfiguration_targetSamplingRate :: Lens' DataPreProcessingConfiguration (Maybe TargetSamplingRate)

The sampling rate of the data after post processing by Amazon Lookout for Equipment. For example, if you provide data that has been collected at a 1 second level and you want the system to resample the data at a 1 minute rate before training, the TargetSamplingRate is 1 minute.

When providing a value for the TargetSamplingRate, you must attach the prefix "PT" to the rate you want. The value for a 1 second rate is therefore PT1S, the value for a 15 minute rate is PT15M, and the value for a 1 hour rate is PT1H

DatasetSchema

data DatasetSchema

Provides information about the data schema used with the given dataset.

See: newDatasetSchema smart constructor.

Constructors

DatasetSchema'

Fields

inlineDataSchema :: Maybe Text

newDatasetSchema :: DatasetSchema

Create a value of DatasetSchema with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:inlineDataSchema:DatasetSchema', datasetSchema_inlineDataSchema -

datasetSchema_inlineDataSchema :: Lens' DatasetSchema (Maybe Text)

DatasetSummary

data DatasetSummary

Contains information about the specific data set, including name, ARN, and status.

See: newDatasetSummary smart constructor.

Constructors

DatasetSummary'

Fields

status :: Maybe DatasetStatus

Indicates the status of the dataset.

datasetArn :: Maybe Text

The Amazon Resource Name (ARN) of the specified dataset.

createdAt :: Maybe POSIX

The time at which the dataset was created in Amazon Lookout for Equipment.

datasetName :: Maybe Text

The name of the dataset.

newDatasetSummary :: DatasetSummary

Create a value of DatasetSummary with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:status:DatasetSummary', datasetSummary_status - Indicates the status of the dataset.

$sel:datasetArn:DatasetSummary', datasetSummary_datasetArn - The Amazon Resource Name (ARN) of the specified dataset.

$sel:createdAt:DatasetSummary', datasetSummary_createdAt - The time at which the dataset was created in Amazon Lookout for Equipment.

$sel:datasetName:DatasetSummary', datasetSummary_datasetName - The name of the dataset.

datasetSummary_status :: Lens' DatasetSummary (Maybe DatasetStatus)

Indicates the status of the dataset.

datasetSummary_datasetArn :: Lens' DatasetSummary (Maybe Text)

The Amazon Resource Name (ARN) of the specified dataset.

datasetSummary_createdAt :: Lens' DatasetSummary (Maybe UTCTime)

The time at which the dataset was created in Amazon Lookout for Equipment.

datasetSummary_datasetName :: Lens' DatasetSummary (Maybe Text)

The name of the dataset.

InferenceExecutionSummary

data InferenceExecutionSummary

Contains information about the specific inference execution, including input and output data configuration, inference scheduling information, status, and so on.

See: newInferenceExecutionSummary smart constructor.

Constructors

InferenceExecutionSummary'

Fields

status :: Maybe InferenceExecutionStatus

Indicates the status of the inference execution.

failedReason :: Maybe Text

Specifies the reason for failure when an inference execution has failed.

modelArn :: Maybe Text

The Amazon Resource Name (ARN) of the ML model used for the inference execution.

dataStartTime :: Maybe POSIX

Indicates the time reference in the dataset at which the inference execution began.

modelName :: Maybe Text

The name of the ML model being used for the inference execution.

customerResultObject :: Maybe S3Object

inferenceSchedulerArn :: Maybe Text

The Amazon Resource Name (ARN) of the inference scheduler being used for the inference execution.

scheduledStartTime :: Maybe POSIX

Indicates the start time at which the inference scheduler began the specific inference execution.

dataOutputConfiguration :: Maybe InferenceOutputConfiguration

Specifies configuration information for the output results from for the inference execution, including the output S3 location.

dataEndTime :: Maybe POSIX

Indicates the time reference in the dataset at which the inference execution stopped.

inferenceSchedulerName :: Maybe Text

The name of the inference scheduler being used for the inference execution.

dataInputConfiguration :: Maybe InferenceInputConfiguration

Specifies configuration information for the input data for the inference scheduler, including delimiter, format, and dataset location.

newInferenceExecutionSummary :: InferenceExecutionSummary

Create a value of InferenceExecutionSummary with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:status:InferenceExecutionSummary', inferenceExecutionSummary_status - Indicates the status of the inference execution.

$sel:failedReason:InferenceExecutionSummary', inferenceExecutionSummary_failedReason - Specifies the reason for failure when an inference execution has failed.

$sel:modelArn:InferenceExecutionSummary', inferenceExecutionSummary_modelArn - The Amazon Resource Name (ARN) of the ML model used for the inference execution.

$sel:dataStartTime:InferenceExecutionSummary', inferenceExecutionSummary_dataStartTime - Indicates the time reference in the dataset at which the inference execution began.

$sel:modelName:InferenceExecutionSummary', inferenceExecutionSummary_modelName - The name of the ML model being used for the inference execution.

$sel:customerResultObject:InferenceExecutionSummary', inferenceExecutionSummary_customerResultObject -

$sel:inferenceSchedulerArn:InferenceExecutionSummary', inferenceExecutionSummary_inferenceSchedulerArn - The Amazon Resource Name (ARN) of the inference scheduler being used for the inference execution.

$sel:scheduledStartTime:InferenceExecutionSummary', inferenceExecutionSummary_scheduledStartTime - Indicates the start time at which the inference scheduler began the specific inference execution.

$sel:dataOutputConfiguration:InferenceExecutionSummary', inferenceExecutionSummary_dataOutputConfiguration - Specifies configuration information for the output results from for the inference execution, including the output S3 location.

$sel:dataEndTime:InferenceExecutionSummary', inferenceExecutionSummary_dataEndTime - Indicates the time reference in the dataset at which the inference execution stopped.

$sel:inferenceSchedulerName:InferenceExecutionSummary', inferenceExecutionSummary_inferenceSchedulerName - The name of the inference scheduler being used for the inference execution.

$sel:dataInputConfiguration:InferenceExecutionSummary', inferenceExecutionSummary_dataInputConfiguration - Specifies configuration information for the input data for the inference scheduler, including delimiter, format, and dataset location.

inferenceExecutionSummary_status :: Lens' InferenceExecutionSummary (Maybe InferenceExecutionStatus)

Indicates the status of the inference execution.

inferenceExecutionSummary_failedReason :: Lens' InferenceExecutionSummary (Maybe Text)

Specifies the reason for failure when an inference execution has failed.

inferenceExecutionSummary_modelArn :: Lens' InferenceExecutionSummary (Maybe Text)

The Amazon Resource Name (ARN) of the ML model used for the inference execution.

inferenceExecutionSummary_dataStartTime :: Lens' InferenceExecutionSummary (Maybe UTCTime)

Indicates the time reference in the dataset at which the inference execution began.

inferenceExecutionSummary_modelName :: Lens' InferenceExecutionSummary (Maybe Text)

The name of the ML model being used for the inference execution.

inferenceExecutionSummary_customerResultObject :: Lens' InferenceExecutionSummary (Maybe S3Object)

inferenceExecutionSummary_inferenceSchedulerArn :: Lens' InferenceExecutionSummary (Maybe Text)

The Amazon Resource Name (ARN) of the inference scheduler being used for the inference execution.

inferenceExecutionSummary_scheduledStartTime :: Lens' InferenceExecutionSummary (Maybe UTCTime)

Indicates the start time at which the inference scheduler began the specific inference execution.

inferenceExecutionSummary_dataOutputConfiguration :: Lens' InferenceExecutionSummary (Maybe InferenceOutputConfiguration)

Specifies configuration information for the output results from for the inference execution, including the output S3 location.

inferenceExecutionSummary_dataEndTime :: Lens' InferenceExecutionSummary (Maybe UTCTime)

Indicates the time reference in the dataset at which the inference execution stopped.

inferenceExecutionSummary_inferenceSchedulerName :: Lens' InferenceExecutionSummary (Maybe Text)

The name of the inference scheduler being used for the inference execution.

inferenceExecutionSummary_dataInputConfiguration :: Lens' InferenceExecutionSummary (Maybe InferenceInputConfiguration)

Specifies configuration information for the input data for the inference scheduler, including delimiter, format, and dataset location.

InferenceInputConfiguration

data InferenceInputConfiguration

Specifies configuration information for the input data for the inference, including S3 location of input data..

See: newInferenceInputConfiguration smart constructor.

Constructors

InferenceInputConfiguration'

Fields

inputTimeZoneOffset :: Maybe Text

Indicates the difference between your time zone and Greenwich Mean Time (GMT).

s3InputConfiguration :: Maybe InferenceS3InputConfiguration

Specifies configuration information for the input data for the inference, including S3 location of input data..

inferenceInputNameConfiguration :: Maybe InferenceInputNameConfiguration

Specifies configuration information for the input data for the inference, including timestamp format and delimiter.

newInferenceInputConfiguration :: InferenceInputConfiguration

Create a value of InferenceInputConfiguration with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:inputTimeZoneOffset:InferenceInputConfiguration', inferenceInputConfiguration_inputTimeZoneOffset - Indicates the difference between your time zone and Greenwich Mean Time (GMT).

$sel:s3InputConfiguration:InferenceInputConfiguration', inferenceInputConfiguration_s3InputConfiguration - Specifies configuration information for the input data for the inference, including S3 location of input data..

$sel:inferenceInputNameConfiguration:InferenceInputConfiguration', inferenceInputConfiguration_inferenceInputNameConfiguration - Specifies configuration information for the input data for the inference, including timestamp format and delimiter.

inferenceInputConfiguration_inputTimeZoneOffset :: Lens' InferenceInputConfiguration (Maybe Text)

Indicates the difference between your time zone and Greenwich Mean Time (GMT).

inferenceInputConfiguration_s3InputConfiguration :: Lens' InferenceInputConfiguration (Maybe InferenceS3InputConfiguration)

Specifies configuration information for the input data for the inference, including S3 location of input data..

inferenceInputConfiguration_inferenceInputNameConfiguration :: Lens' InferenceInputConfiguration (Maybe InferenceInputNameConfiguration)

Specifies configuration information for the input data for the inference, including timestamp format and delimiter.

InferenceInputNameConfiguration

data InferenceInputNameConfiguration

Specifies configuration information for the input data for the inference, including timestamp format and delimiter.

See: newInferenceInputNameConfiguration smart constructor.

Constructors

InferenceInputNameConfiguration'

Fields

timestampFormat :: Maybe Text

The format of the timestamp, whether Epoch time, or standard, with or without hyphens (-).

componentTimestampDelimiter :: Maybe Text

Indicates the delimiter character used between items in the data.

newInferenceInputNameConfiguration :: InferenceInputNameConfiguration

Create a value of InferenceInputNameConfiguration with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:timestampFormat:InferenceInputNameConfiguration', inferenceInputNameConfiguration_timestampFormat - The format of the timestamp, whether Epoch time, or standard, with or without hyphens (-).

$sel:componentTimestampDelimiter:InferenceInputNameConfiguration', inferenceInputNameConfiguration_componentTimestampDelimiter - Indicates the delimiter character used between items in the data.

inferenceInputNameConfiguration_timestampFormat :: Lens' InferenceInputNameConfiguration (Maybe Text)

The format of the timestamp, whether Epoch time, or standard, with or without hyphens (-).

inferenceInputNameConfiguration_componentTimestampDelimiter :: Lens' InferenceInputNameConfiguration (Maybe Text)

Indicates the delimiter character used between items in the data.

InferenceOutputConfiguration

data InferenceOutputConfiguration

Specifies configuration information for the output results from for the inference, including KMS key ID and output S3 location.

See: newInferenceOutputConfiguration smart constructor.

Constructors

InferenceOutputConfiguration'

Fields

kmsKeyId :: Maybe Text

The ID number for the AWS KMS key used to encrypt the inference output.

s3OutputConfiguration :: InferenceS3OutputConfiguration

Specifies configuration information for the output results from for the inference, output S3 location.

newInferenceOutputConfiguration

Arguments

:: InferenceS3OutputConfiguration
$sel:s3OutputConfiguration:InferenceOutputConfiguration'

-> InferenceOutputConfiguration

Create a value of InferenceOutputConfiguration with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:kmsKeyId:InferenceOutputConfiguration', inferenceOutputConfiguration_kmsKeyId - The ID number for the AWS KMS key used to encrypt the inference output.

$sel:s3OutputConfiguration:InferenceOutputConfiguration', inferenceOutputConfiguration_s3OutputConfiguration - Specifies configuration information for the output results from for the inference, output S3 location.

inferenceOutputConfiguration_kmsKeyId :: Lens' InferenceOutputConfiguration (Maybe Text)

The ID number for the AWS KMS key used to encrypt the inference output.

inferenceOutputConfiguration_s3OutputConfiguration :: Lens' InferenceOutputConfiguration InferenceS3OutputConfiguration

Specifies configuration information for the output results from for the inference, output S3 location.

InferenceS3InputConfiguration

data InferenceS3InputConfiguration

Specifies configuration information for the input data for the inference, including input data S3 location.

See: newInferenceS3InputConfiguration smart constructor.

Constructors

InferenceS3InputConfiguration'

Fields

prefix :: Maybe Text

The prefix for the S3 bucket used for the input data for the inference.

bucket :: Text

The bucket containing the input dataset for the inference.

newInferenceS3InputConfiguration

Arguments

:: Text
$sel:bucket:InferenceS3InputConfiguration'

-> InferenceS3InputConfiguration

Create a value of InferenceS3InputConfiguration with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:prefix:InferenceS3InputConfiguration', inferenceS3InputConfiguration_prefix - The prefix for the S3 bucket used for the input data for the inference.

$sel:bucket:InferenceS3InputConfiguration', inferenceS3InputConfiguration_bucket - The bucket containing the input dataset for the inference.

inferenceS3InputConfiguration_prefix :: Lens' InferenceS3InputConfiguration (Maybe Text)

The prefix for the S3 bucket used for the input data for the inference.

inferenceS3InputConfiguration_bucket :: Lens' InferenceS3InputConfiguration Text

The bucket containing the input dataset for the inference.

InferenceS3OutputConfiguration

data InferenceS3OutputConfiguration

Specifies configuration information for the output results from the inference, including output S3 location.

See: newInferenceS3OutputConfiguration smart constructor.

Constructors

InferenceS3OutputConfiguration'

Fields

prefix :: Maybe Text

The prefix for the S3 bucket used for the output results from the inference.

bucket :: Text

The bucket containing the output results from the inference

newInferenceS3OutputConfiguration

Arguments

:: Text
$sel:bucket:InferenceS3OutputConfiguration'

-> InferenceS3OutputConfiguration

Create a value of InferenceS3OutputConfiguration with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:prefix:InferenceS3OutputConfiguration', inferenceS3OutputConfiguration_prefix - The prefix for the S3 bucket used for the output results from the inference.

$sel:bucket:InferenceS3OutputConfiguration', inferenceS3OutputConfiguration_bucket - The bucket containing the output results from the inference

inferenceS3OutputConfiguration_prefix :: Lens' InferenceS3OutputConfiguration (Maybe Text)

The prefix for the S3 bucket used for the output results from the inference.

inferenceS3OutputConfiguration_bucket :: Lens' InferenceS3OutputConfiguration Text

The bucket containing the output results from the inference

InferenceSchedulerSummary

data InferenceSchedulerSummary

Contains information about the specific inference scheduler, including data delay offset, model name and ARN, status, and so on.

See: newInferenceSchedulerSummary smart constructor.

Constructors

InferenceSchedulerSummary'

Fields

status :: Maybe InferenceSchedulerStatus

Indicates the status of the inference scheduler.

dataUploadFrequency :: Maybe DataUploadFrequency

How often data is uploaded to the source S3 bucket for the input data. This value is the length of time between data uploads. For instance, if you select 5 minutes, Amazon Lookout for Equipment will upload the real-time data to the source bucket once every 5 minutes. This frequency also determines how often Amazon Lookout for Equipment starts a scheduled inference on your data. In this example, it starts once every 5 minutes.

dataDelayOffsetInMinutes :: Maybe Natural

A period of time (in minutes) by which inference on the data is delayed after the data starts. For instance, if an offset delay time of five minutes was selected, inference will not begin on the data until the first data measurement after the five minute mark. For example, if five minutes is selected, the inference scheduler will wake up at the configured frequency with the additional five minute delay time to check the customer S3 bucket. The customer can upload data at the same frequency and they don't need to stop and restart the scheduler when uploading new data.

modelArn :: Maybe Text

The Amazon Resource Name (ARN) of the ML model used by the inference scheduler.

modelName :: Maybe Text

The name of the ML model used for the inference scheduler.

inferenceSchedulerArn :: Maybe Text

The Amazon Resource Name (ARN) of the inference scheduler.

inferenceSchedulerName :: Maybe Text

The name of the inference scheduler.

newInferenceSchedulerSummary :: InferenceSchedulerSummary

Create a value of InferenceSchedulerSummary with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:status:InferenceSchedulerSummary', inferenceSchedulerSummary_status - Indicates the status of the inference scheduler.

$sel:dataUploadFrequency:InferenceSchedulerSummary', inferenceSchedulerSummary_dataUploadFrequency - How often data is uploaded to the source S3 bucket for the input data. This value is the length of time between data uploads. For instance, if you select 5 minutes, Amazon Lookout for Equipment will upload the real-time data to the source bucket once every 5 minutes. This frequency also determines how often Amazon Lookout for Equipment starts a scheduled inference on your data. In this example, it starts once every 5 minutes.

$sel:dataDelayOffsetInMinutes:InferenceSchedulerSummary', inferenceSchedulerSummary_dataDelayOffsetInMinutes - A period of time (in minutes) by which inference on the data is delayed after the data starts. For instance, if an offset delay time of five minutes was selected, inference will not begin on the data until the first data measurement after the five minute mark. For example, if five minutes is selected, the inference scheduler will wake up at the configured frequency with the additional five minute delay time to check the customer S3 bucket. The customer can upload data at the same frequency and they don't need to stop and restart the scheduler when uploading new data.

$sel:modelArn:InferenceSchedulerSummary', inferenceSchedulerSummary_modelArn - The Amazon Resource Name (ARN) of the ML model used by the inference scheduler.

$sel:modelName:InferenceSchedulerSummary', inferenceSchedulerSummary_modelName - The name of the ML model used for the inference scheduler.

$sel:inferenceSchedulerArn:InferenceSchedulerSummary', inferenceSchedulerSummary_inferenceSchedulerArn - The Amazon Resource Name (ARN) of the inference scheduler.

$sel:inferenceSchedulerName:InferenceSchedulerSummary', inferenceSchedulerSummary_inferenceSchedulerName - The name of the inference scheduler.

inferenceSchedulerSummary_status :: Lens' InferenceSchedulerSummary (Maybe InferenceSchedulerStatus)

Indicates the status of the inference scheduler.

inferenceSchedulerSummary_dataUploadFrequency :: Lens' InferenceSchedulerSummary (Maybe DataUploadFrequency)

How often data is uploaded to the source S3 bucket for the input data. This value is the length of time between data uploads. For instance, if you select 5 minutes, Amazon Lookout for Equipment will upload the real-time data to the source bucket once every 5 minutes. This frequency also determines how often Amazon Lookout for Equipment starts a scheduled inference on your data. In this example, it starts once every 5 minutes.

inferenceSchedulerSummary_dataDelayOffsetInMinutes :: Lens' InferenceSchedulerSummary (Maybe Natural)

A period of time (in minutes) by which inference on the data is delayed after the data starts. For instance, if an offset delay time of five minutes was selected, inference will not begin on the data until the first data measurement after the five minute mark. For example, if five minutes is selected, the inference scheduler will wake up at the configured frequency with the additional five minute delay time to check the customer S3 bucket. The customer can upload data at the same frequency and they don't need to stop and restart the scheduler when uploading new data.

inferenceSchedulerSummary_modelArn :: Lens' InferenceSchedulerSummary (Maybe Text)

The Amazon Resource Name (ARN) of the ML model used by the inference scheduler.

inferenceSchedulerSummary_modelName :: Lens' InferenceSchedulerSummary (Maybe Text)

The name of the ML model used for the inference scheduler.

inferenceSchedulerSummary_inferenceSchedulerArn :: Lens' InferenceSchedulerSummary (Maybe Text)

The Amazon Resource Name (ARN) of the inference scheduler.

inferenceSchedulerSummary_inferenceSchedulerName :: Lens' InferenceSchedulerSummary (Maybe Text)

The name of the inference scheduler.

IngestionInputConfiguration

data IngestionInputConfiguration

Specifies configuration information for the input data for the data ingestion job, including input data S3 location.

See: newIngestionInputConfiguration smart constructor.

Constructors

IngestionInputConfiguration'

Fields

s3InputConfiguration :: IngestionS3InputConfiguration

The location information for the S3 bucket used for input data for the data ingestion.

newIngestionInputConfiguration

Arguments

:: IngestionS3InputConfiguration
$sel:s3InputConfiguration:IngestionInputConfiguration'

-> IngestionInputConfiguration

Create a value of IngestionInputConfiguration with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:s3InputConfiguration:IngestionInputConfiguration', ingestionInputConfiguration_s3InputConfiguration - The location information for the S3 bucket used for input data for the data ingestion.

ingestionInputConfiguration_s3InputConfiguration :: Lens' IngestionInputConfiguration IngestionS3InputConfiguration

The location information for the S3 bucket used for input data for the data ingestion.

IngestionS3InputConfiguration

data IngestionS3InputConfiguration

Specifies S3 configuration information for the input data for the data ingestion job.

See: newIngestionS3InputConfiguration smart constructor.

Constructors

IngestionS3InputConfiguration'

Fields

prefix :: Maybe Text

The prefix for the S3 location being used for the input data for the data ingestion.

bucket :: Text

The name of the S3 bucket used for the input data for the data ingestion.

newIngestionS3InputConfiguration

Arguments

:: Text
$sel:bucket:IngestionS3InputConfiguration'

-> IngestionS3InputConfiguration

Create a value of IngestionS3InputConfiguration with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:prefix:IngestionS3InputConfiguration', ingestionS3InputConfiguration_prefix - The prefix for the S3 location being used for the input data for the data ingestion.

$sel:bucket:IngestionS3InputConfiguration', ingestionS3InputConfiguration_bucket - The name of the S3 bucket used for the input data for the data ingestion.

ingestionS3InputConfiguration_prefix :: Lens' IngestionS3InputConfiguration (Maybe Text)

The prefix for the S3 location being used for the input data for the data ingestion.

ingestionS3InputConfiguration_bucket :: Lens' IngestionS3InputConfiguration Text

The name of the S3 bucket used for the input data for the data ingestion.

LabelsInputConfiguration

data LabelsInputConfiguration

Contains the configuration information for the S3 location being used to hold label data.

See: newLabelsInputConfiguration smart constructor.

Constructors

LabelsInputConfiguration'

Fields

s3InputConfiguration :: LabelsS3InputConfiguration

Contains location information for the S3 location being used for label data.

newLabelsInputConfiguration

Arguments

:: LabelsS3InputConfiguration
$sel:s3InputConfiguration:LabelsInputConfiguration'

-> LabelsInputConfiguration

Create a value of LabelsInputConfiguration with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:s3InputConfiguration:LabelsInputConfiguration', labelsInputConfiguration_s3InputConfiguration - Contains location information for the S3 location being used for label data.

labelsInputConfiguration_s3InputConfiguration :: Lens' LabelsInputConfiguration LabelsS3InputConfiguration

Contains location information for the S3 location being used for label data.

LabelsS3InputConfiguration

data LabelsS3InputConfiguration

The location information (prefix and bucket name) for the s3 location being used for label data.

See: newLabelsS3InputConfiguration smart constructor.

Constructors

LabelsS3InputConfiguration'

Fields

prefix :: Maybe Text

The prefix for the S3 bucket used for the label data.

bucket :: Text

The name of the S3 bucket holding the label data.

newLabelsS3InputConfiguration

Arguments

:: Text
$sel:bucket:LabelsS3InputConfiguration'

-> LabelsS3InputConfiguration

Create a value of LabelsS3InputConfiguration with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:prefix:LabelsS3InputConfiguration', labelsS3InputConfiguration_prefix - The prefix for the S3 bucket used for the label data.

$sel:bucket:LabelsS3InputConfiguration', labelsS3InputConfiguration_bucket - The name of the S3 bucket holding the label data.

labelsS3InputConfiguration_prefix :: Lens' LabelsS3InputConfiguration (Maybe Text)

The prefix for the S3 bucket used for the label data.

labelsS3InputConfiguration_bucket :: Lens' LabelsS3InputConfiguration Text

The name of the S3 bucket holding the label data.

ModelSummary

data ModelSummary

Provides information about the specified ML model, including dataset and model names and ARNs, as well as status.

See: newModelSummary smart constructor.

Constructors

ModelSummary'

Fields

status :: Maybe ModelStatus

Indicates the status of the ML model.

datasetArn :: Maybe Text

The Amazon Resource Name (ARN) of the dataset used to create the model.

modelArn :: Maybe Text

The Amazon Resource Name (ARN) of the ML model.

createdAt :: Maybe POSIX

The time at which the specific model was created.

modelName :: Maybe Text

The name of the ML model.

datasetName :: Maybe Text

The name of the dataset being used for the ML model.

newModelSummary :: ModelSummary

Create a value of ModelSummary with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:status:ModelSummary', modelSummary_status - Indicates the status of the ML model.

$sel:datasetArn:ModelSummary', modelSummary_datasetArn - The Amazon Resource Name (ARN) of the dataset used to create the model.

$sel:modelArn:ModelSummary', modelSummary_modelArn - The Amazon Resource Name (ARN) of the ML model.

$sel:createdAt:ModelSummary', modelSummary_createdAt - The time at which the specific model was created.

$sel:modelName:ModelSummary', modelSummary_modelName - The name of the ML model.

$sel:datasetName:ModelSummary', modelSummary_datasetName - The name of the dataset being used for the ML model.

modelSummary_status :: Lens' ModelSummary (Maybe ModelStatus)

Indicates the status of the ML model.

modelSummary_datasetArn :: Lens' ModelSummary (Maybe Text)

The Amazon Resource Name (ARN) of the dataset used to create the model.

modelSummary_modelArn :: Lens' ModelSummary (Maybe Text)

The Amazon Resource Name (ARN) of the ML model.

modelSummary_createdAt :: Lens' ModelSummary (Maybe UTCTime)

The time at which the specific model was created.

modelSummary_modelName :: Lens' ModelSummary (Maybe Text)

The name of the ML model.

modelSummary_datasetName :: Lens' ModelSummary (Maybe Text)

The name of the dataset being used for the ML model.

S3Object

data S3Object

Contains information about an S3 bucket.

See: newS3Object smart constructor.

Constructors

S3Object'

Fields

bucket :: Text

The name of the specific S3 bucket.

key :: Text

The AWS Key Management Service (AWS KMS) key being used to encrypt the S3 object. Without this key, data in the bucket is not accessible.

newS3Object

Arguments

:: Text
$sel:bucket:S3Object'

-> Text
$sel:key:S3Object'

-> S3Object

Create a value of S3Object with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:bucket:S3Object', s3Object_bucket - The name of the specific S3 bucket.

$sel:key:S3Object', s3Object_key - The AWS Key Management Service (AWS KMS) key being used to encrypt the S3 object. Without this key, data in the bucket is not accessible.

s3Object_bucket :: Lens' S3Object Text

The name of the specific S3 bucket.

s3Object_key :: Lens' S3Object Text

The AWS Key Management Service (AWS KMS) key being used to encrypt the S3 object. Without this key, data in the bucket is not accessible.

Tag

data Tag

A tag is a key-value pair that can be added to a resource as metadata.

See: newTag smart constructor.

Constructors

Tag'

Fields

key :: Text

The key for the specified tag.

value :: Text

The value for the specified tag.

newTag

Arguments

:: Text
$sel:key:Tag'

-> Text
$sel:value:Tag'

-> Tag

Create a value of Tag with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:key:Tag', tag_key - The key for the specified tag.

$sel:value:Tag', tag_value - The value for the specified tag.

tag_key :: Lens' Tag Text

The key for the specified tag.

tag_value :: Lens' Tag Text

The value for the specified tag.