Amazonka.LookoutMetrics.CreateMetricSet

Description

Creates a dataset.

Creating a Request

data CreateMetricSet

See: newCreateMetricSet smart constructor.

Constructors

CreateMetricSet'

Fields

dimensionList :: Maybe (NonEmpty Text)

A list of the fields you want to treat as dimensions.

offset :: Maybe Natural

After an interval ends, the amount of seconds that the detector waits before importing data. Offset is only supported for S3 and Redshift datasources.

timestampColumn :: Maybe TimestampColumn

Contains information about the column used for tracking time in your source data.

metricSetFrequency :: Maybe Frequency

The frequency with which the source data will be analyzed for anomalies.

metricSetDescription :: Maybe Text

A description of the dataset you are creating.

timezone :: Maybe Text

The time zone in which your source data was recorded.

tags :: Maybe (HashMap Text Text)

A list of tags to apply to the dataset.

anomalyDetectorArn :: Text

The ARN of the anomaly detector that will use the dataset.

metricSetName :: Text

The name of the dataset.

metricList :: NonEmpty Metric

A list of metrics that the dataset will contain.

metricSource :: MetricSource

Contains information about how the source data should be interpreted.

newCreateMetricSet

Arguments

:: Text
$sel:anomalyDetectorArn:CreateMetricSet'

-> Text
$sel:metricSetName:CreateMetricSet'

-> NonEmpty Metric
$sel:metricList:CreateMetricSet'

-> MetricSource
$sel:metricSource:CreateMetricSet'

-> CreateMetricSet

Create a value of CreateMetricSet with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:dimensionList:CreateMetricSet', createMetricSet_dimensionList - A list of the fields you want to treat as dimensions.

$sel:offset:CreateMetricSet', createMetricSet_offset - After an interval ends, the amount of seconds that the detector waits before importing data. Offset is only supported for S3 and Redshift datasources.

$sel:timestampColumn:CreateMetricSet', createMetricSet_timestampColumn - Contains information about the column used for tracking time in your source data.

$sel:metricSetFrequency:CreateMetricSet', createMetricSet_metricSetFrequency - The frequency with which the source data will be analyzed for anomalies.

$sel:metricSetDescription:CreateMetricSet', createMetricSet_metricSetDescription - A description of the dataset you are creating.

$sel:timezone:CreateMetricSet', createMetricSet_timezone - The time zone in which your source data was recorded.

$sel:tags:CreateMetricSet', createMetricSet_tags - A list of tags to apply to the dataset.

$sel:anomalyDetectorArn:CreateMetricSet', createMetricSet_anomalyDetectorArn - The ARN of the anomaly detector that will use the dataset.

$sel:metricSetName:CreateMetricSet', createMetricSet_metricSetName - The name of the dataset.

$sel:metricList:CreateMetricSet', createMetricSet_metricList - A list of metrics that the dataset will contain.

$sel:metricSource:CreateMetricSet', createMetricSet_metricSource - Contains information about how the source data should be interpreted.

Request Lenses

createMetricSet_dimensionList :: Lens' CreateMetricSet (Maybe (NonEmpty Text))

A list of the fields you want to treat as dimensions.

createMetricSet_offset :: Lens' CreateMetricSet (Maybe Natural)

After an interval ends, the amount of seconds that the detector waits before importing data. Offset is only supported for S3 and Redshift datasources.

createMetricSet_timestampColumn :: Lens' CreateMetricSet (Maybe TimestampColumn)

Contains information about the column used for tracking time in your source data.

createMetricSet_metricSetFrequency :: Lens' CreateMetricSet (Maybe Frequency)

The frequency with which the source data will be analyzed for anomalies.

createMetricSet_metricSetDescription :: Lens' CreateMetricSet (Maybe Text)

A description of the dataset you are creating.

createMetricSet_timezone :: Lens' CreateMetricSet (Maybe Text)

The time zone in which your source data was recorded.

createMetricSet_tags :: Lens' CreateMetricSet (Maybe (HashMap Text Text))

A list of tags to apply to the dataset.

createMetricSet_anomalyDetectorArn :: Lens' CreateMetricSet Text

The ARN of the anomaly detector that will use the dataset.

createMetricSet_metricSetName :: Lens' CreateMetricSet Text

The name of the dataset.

createMetricSet_metricList :: Lens' CreateMetricSet (NonEmpty Metric)

A list of metrics that the dataset will contain.

createMetricSet_metricSource :: Lens' CreateMetricSet MetricSource

Contains information about how the source data should be interpreted.

Destructuring the Response

data CreateMetricSetResponse

See: newCreateMetricSetResponse smart constructor.

Constructors

CreateMetricSetResponse'

Fields

metricSetArn :: Maybe Text

The ARN of the dataset.

httpStatus :: Int

The response's http status code.

newCreateMetricSetResponse

Arguments

:: Int
$sel:httpStatus:CreateMetricSetResponse'

-> CreateMetricSetResponse

Create a value of CreateMetricSetResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:metricSetArn:CreateMetricSetResponse', createMetricSetResponse_metricSetArn - The ARN of the dataset.

$sel:httpStatus:CreateMetricSetResponse', createMetricSetResponse_httpStatus - The response's http status code.

Response Lenses

createMetricSetResponse_metricSetArn :: Lens' CreateMetricSetResponse (Maybe Text)

The ARN of the dataset.

createMetricSetResponse_httpStatus :: Lens' CreateMetricSetResponse Int

The response's http status code.