Amazonka.LookoutMetrics.DescribeAnomalyDetector

Description

Describes a detector.

Amazon Lookout for Metrics API actions are eventually consistent. If you do a read operation on a resource immediately after creating or modifying it, use retries to allow time for the write operation to complete.

Creating a Request

data DescribeAnomalyDetector

See: newDescribeAnomalyDetector smart constructor.

Constructors

DescribeAnomalyDetector'

Fields

anomalyDetectorArn :: Text

The ARN of the detector to describe.

newDescribeAnomalyDetector

Arguments

:: Text
$sel:anomalyDetectorArn:DescribeAnomalyDetector'

-> DescribeAnomalyDetector

Create a value of DescribeAnomalyDetector with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:anomalyDetectorArn:DescribeAnomalyDetector', describeAnomalyDetector_anomalyDetectorArn - The ARN of the detector to describe.

Request Lenses

describeAnomalyDetector_anomalyDetectorArn :: Lens' DescribeAnomalyDetector Text

The ARN of the detector to describe.

Destructuring the Response

data DescribeAnomalyDetectorResponse

See: newDescribeAnomalyDetectorResponse smart constructor.

Constructors

DescribeAnomalyDetectorResponse'

Fields

creationTime :: Maybe POSIX

The time at which the detector was created.

status :: Maybe AnomalyDetectorStatus

The status of the detector.

failureReason :: Maybe Text

The reason that the detector failed, if any.

kmsKeyArn :: Maybe Text

The ARN of the KMS key to use to encrypt your data.

anomalyDetectorArn :: Maybe Text

The ARN of the detector.

anomalyDetectorConfig :: Maybe AnomalyDetectorConfigSummary

Contains information about the detector's configuration.

anomalyDetectorName :: Maybe Text

The name of the detector.

anomalyDetectorDescription :: Maybe Text

A description of the detector.

lastModificationTime :: Maybe POSIX

The time at which the detector was last modified.

httpStatus :: Int

The response's http status code.

newDescribeAnomalyDetectorResponse

Arguments

:: Int
$sel:httpStatus:DescribeAnomalyDetectorResponse'

-> DescribeAnomalyDetectorResponse

Create a value of DescribeAnomalyDetectorResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:creationTime:DescribeAnomalyDetectorResponse', describeAnomalyDetectorResponse_creationTime - The time at which the detector was created.

$sel:status:DescribeAnomalyDetectorResponse', describeAnomalyDetectorResponse_status - The status of the detector.

$sel:failureReason:DescribeAnomalyDetectorResponse', describeAnomalyDetectorResponse_failureReason - The reason that the detector failed, if any.

$sel:kmsKeyArn:DescribeAnomalyDetectorResponse', describeAnomalyDetectorResponse_kmsKeyArn - The ARN of the KMS key to use to encrypt your data.

$sel:anomalyDetectorArn:DescribeAnomalyDetector', describeAnomalyDetectorResponse_anomalyDetectorArn - The ARN of the detector.

$sel:anomalyDetectorConfig:DescribeAnomalyDetectorResponse', describeAnomalyDetectorResponse_anomalyDetectorConfig - Contains information about the detector's configuration.

$sel:anomalyDetectorName:DescribeAnomalyDetectorResponse', describeAnomalyDetectorResponse_anomalyDetectorName - The name of the detector.

$sel:anomalyDetectorDescription:DescribeAnomalyDetectorResponse', describeAnomalyDetectorResponse_anomalyDetectorDescription - A description of the detector.

$sel:lastModificationTime:DescribeAnomalyDetectorResponse', describeAnomalyDetectorResponse_lastModificationTime - The time at which the detector was last modified.

$sel:httpStatus:DescribeAnomalyDetectorResponse', describeAnomalyDetectorResponse_httpStatus - The response's http status code.

Response Lenses

describeAnomalyDetectorResponse_creationTime :: Lens' DescribeAnomalyDetectorResponse (Maybe UTCTime)

The time at which the detector was created.

describeAnomalyDetectorResponse_status :: Lens' DescribeAnomalyDetectorResponse (Maybe AnomalyDetectorStatus)

The status of the detector.

describeAnomalyDetectorResponse_failureReason :: Lens' DescribeAnomalyDetectorResponse (Maybe Text)

The reason that the detector failed, if any.

describeAnomalyDetectorResponse_kmsKeyArn :: Lens' DescribeAnomalyDetectorResponse (Maybe Text)

The ARN of the KMS key to use to encrypt your data.

describeAnomalyDetectorResponse_anomalyDetectorArn :: Lens' DescribeAnomalyDetectorResponse (Maybe Text)

The ARN of the detector.

describeAnomalyDetectorResponse_anomalyDetectorConfig :: Lens' DescribeAnomalyDetectorResponse (Maybe AnomalyDetectorConfigSummary)

Contains information about the detector's configuration.

describeAnomalyDetectorResponse_anomalyDetectorName :: Lens' DescribeAnomalyDetectorResponse (Maybe Text)

The name of the detector.

describeAnomalyDetectorResponse_anomalyDetectorDescription :: Lens' DescribeAnomalyDetectorResponse (Maybe Text)

A description of the detector.

describeAnomalyDetectorResponse_lastModificationTime :: Lens' DescribeAnomalyDetectorResponse (Maybe UTCTime)

The time at which the detector was last modified.

describeAnomalyDetectorResponse_httpStatus :: Lens' DescribeAnomalyDetectorResponse Int

The response's http status code.