Amazonka.LookoutMetrics.ListAnomalyGroupTimeSeries

Description

Gets a list of anomalous metrics for a measure in an anomaly group.

Creating a Request

data ListAnomalyGroupTimeSeries

See: newListAnomalyGroupTimeSeries smart constructor.

Constructors

ListAnomalyGroupTimeSeries'

Fields

nextToken :: Maybe Text

Specify the pagination token that's returned by a previous request to retrieve the next page of results.

maxResults :: Maybe Natural

The maximum number of results to return.

anomalyDetectorArn :: Text

The Amazon Resource Name (ARN) of the anomaly detector.

anomalyGroupId :: Text

The ID of the anomaly group.

metricName :: Text

The name of the measure field.

newListAnomalyGroupTimeSeries

Arguments

:: Text
$sel:anomalyDetectorArn:ListAnomalyGroupTimeSeries'

-> Text
$sel:anomalyGroupId:ListAnomalyGroupTimeSeries'

-> Text
$sel:metricName:ListAnomalyGroupTimeSeries'

-> ListAnomalyGroupTimeSeries

Create a value of ListAnomalyGroupTimeSeries with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:nextToken:ListAnomalyGroupTimeSeries', listAnomalyGroupTimeSeries_nextToken - Specify the pagination token that's returned by a previous request to retrieve the next page of results.

$sel:maxResults:ListAnomalyGroupTimeSeries', listAnomalyGroupTimeSeries_maxResults - The maximum number of results to return.

$sel:anomalyDetectorArn:ListAnomalyGroupTimeSeries', listAnomalyGroupTimeSeries_anomalyDetectorArn - The Amazon Resource Name (ARN) of the anomaly detector.

$sel:anomalyGroupId:ListAnomalyGroupTimeSeries', listAnomalyGroupTimeSeries_anomalyGroupId - The ID of the anomaly group.

$sel:metricName:ListAnomalyGroupTimeSeries', listAnomalyGroupTimeSeries_metricName - The name of the measure field.

Request Lenses

listAnomalyGroupTimeSeries_nextToken :: Lens' ListAnomalyGroupTimeSeries (Maybe Text)

Specify the pagination token that's returned by a previous request to retrieve the next page of results.

listAnomalyGroupTimeSeries_maxResults :: Lens' ListAnomalyGroupTimeSeries (Maybe Natural)

The maximum number of results to return.

listAnomalyGroupTimeSeries_anomalyDetectorArn :: Lens' ListAnomalyGroupTimeSeries Text

The Amazon Resource Name (ARN) of the anomaly detector.

listAnomalyGroupTimeSeries_anomalyGroupId :: Lens' ListAnomalyGroupTimeSeries Text

The ID of the anomaly group.

listAnomalyGroupTimeSeries_metricName :: Lens' ListAnomalyGroupTimeSeries Text

The name of the measure field.

Destructuring the Response

data ListAnomalyGroupTimeSeriesResponse

See: newListAnomalyGroupTimeSeriesResponse smart constructor.

Constructors

ListAnomalyGroupTimeSeriesResponse'

Fields

timeSeriesList :: Maybe [TimeSeries]

A list of anomalous metrics.

timestampList :: Maybe [Text]

Timestamps for the anomalous metrics.

metricName :: Maybe Text

The name of the measure field.

anomalyGroupId :: Maybe Text

The ID of the anomaly group.

nextToken :: Maybe Text

The pagination token that's included if more results are available.

httpStatus :: Int

The response's http status code.

newListAnomalyGroupTimeSeriesResponse

Arguments

:: Int
$sel:httpStatus:ListAnomalyGroupTimeSeriesResponse'

-> ListAnomalyGroupTimeSeriesResponse

Create a value of ListAnomalyGroupTimeSeriesResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:timeSeriesList:ListAnomalyGroupTimeSeriesResponse', listAnomalyGroupTimeSeriesResponse_timeSeriesList - A list of anomalous metrics.

$sel:timestampList:ListAnomalyGroupTimeSeriesResponse', listAnomalyGroupTimeSeriesResponse_timestampList - Timestamps for the anomalous metrics.

$sel:metricName:ListAnomalyGroupTimeSeries', listAnomalyGroupTimeSeriesResponse_metricName - The name of the measure field.

$sel:anomalyGroupId:ListAnomalyGroupTimeSeries', listAnomalyGroupTimeSeriesResponse_anomalyGroupId - The ID of the anomaly group.

$sel:nextToken:ListAnomalyGroupTimeSeries', listAnomalyGroupTimeSeriesResponse_nextToken - The pagination token that's included if more results are available.

$sel:httpStatus:ListAnomalyGroupTimeSeriesResponse', listAnomalyGroupTimeSeriesResponse_httpStatus - The response's http status code.

Response Lenses

listAnomalyGroupTimeSeriesResponse_timeSeriesList :: Lens' ListAnomalyGroupTimeSeriesResponse (Maybe [TimeSeries])

A list of anomalous metrics.

listAnomalyGroupTimeSeriesResponse_timestampList :: Lens' ListAnomalyGroupTimeSeriesResponse (Maybe [Text])

Timestamps for the anomalous metrics.

listAnomalyGroupTimeSeriesResponse_metricName :: Lens' ListAnomalyGroupTimeSeriesResponse (Maybe Text)

The name of the measure field.

listAnomalyGroupTimeSeriesResponse_anomalyGroupId :: Lens' ListAnomalyGroupTimeSeriesResponse (Maybe Text)

The ID of the anomaly group.

listAnomalyGroupTimeSeriesResponse_nextToken :: Lens' ListAnomalyGroupTimeSeriesResponse (Maybe Text)

The pagination token that's included if more results are available.

listAnomalyGroupTimeSeriesResponse_httpStatus :: Lens' ListAnomalyGroupTimeSeriesResponse Int

The response's http status code.