Amazonka.LookoutMetrics.Types.S3SourceConfig

Description

Documentation

data S3SourceConfig

Contains information about the configuration of the S3 bucket that contains source files.

See: newS3SourceConfig smart constructor.

Constructors

S3SourceConfig'

Fields

templatedPathList :: Maybe (NonEmpty Text)

A list of templated paths to the source files.

historicalDataPathList :: Maybe (NonEmpty Text)

A list of paths to the historical data files.

fileFormatDescriptor :: Maybe FileFormatDescriptor

Contains information about a source file's formatting.

roleArn :: Text

The ARN of an IAM role that has read and write access permissions to the source S3 bucket.

newS3SourceConfig

Arguments

:: Text
$sel:roleArn:S3SourceConfig'

-> S3SourceConfig

Create a value of S3SourceConfig with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:templatedPathList:S3SourceConfig', s3SourceConfig_templatedPathList - A list of templated paths to the source files.

$sel:historicalDataPathList:S3SourceConfig', s3SourceConfig_historicalDataPathList - A list of paths to the historical data files.

$sel:fileFormatDescriptor:S3SourceConfig', s3SourceConfig_fileFormatDescriptor - Contains information about a source file's formatting.

$sel:roleArn:S3SourceConfig', s3SourceConfig_roleArn - The ARN of an IAM role that has read and write access permissions to the source S3 bucket.

s3SourceConfig_templatedPathList :: Lens' S3SourceConfig (Maybe (NonEmpty Text))

A list of templated paths to the source files.

s3SourceConfig_historicalDataPathList :: Lens' S3SourceConfig (Maybe (NonEmpty Text))

A list of paths to the historical data files.

s3SourceConfig_fileFormatDescriptor :: Lens' S3SourceConfig (Maybe FileFormatDescriptor)

Contains information about a source file's formatting.

s3SourceConfig_roleArn :: Lens' S3SourceConfig Text

The ARN of an IAM role that has read and write access permissions to the source S3 bucket.