Amazonka.LookoutMetrics.UpdateMetricSet

Description

Updates a dataset.

Creating a Request

data UpdateMetricSet

See: newUpdateMetricSet smart constructor.

Constructors

UpdateMetricSet'

Fields

dimensionList :: Maybe (NonEmpty Text)

The dimension list.

offset :: Maybe Natural

After an interval ends, the amount of seconds that the detector waits before importing data. Offset is only supported for S3 and Redshift datasources.

timestampColumn :: Maybe TimestampColumn

The timestamp column.

metricList :: Maybe (NonEmpty Metric)

The metric list.

metricSource :: Maybe MetricSource

metricSetFrequency :: Maybe Frequency

The dataset's interval.

metricSetDescription :: Maybe Text

The dataset's description.

metricSetArn :: Text

The ARN of the dataset to update.

newUpdateMetricSet

Arguments

:: Text
$sel:metricSetArn:UpdateMetricSet'

-> UpdateMetricSet

Create a value of UpdateMetricSet with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:dimensionList:UpdateMetricSet', updateMetricSet_dimensionList - The dimension list.

$sel:offset:UpdateMetricSet', updateMetricSet_offset - After an interval ends, the amount of seconds that the detector waits before importing data. Offset is only supported for S3 and Redshift datasources.

$sel:timestampColumn:UpdateMetricSet', updateMetricSet_timestampColumn - The timestamp column.

$sel:metricList:UpdateMetricSet', updateMetricSet_metricList - The metric list.

$sel:metricSource:UpdateMetricSet', updateMetricSet_metricSource - Undocumented member.

$sel:metricSetFrequency:UpdateMetricSet', updateMetricSet_metricSetFrequency - The dataset's interval.

$sel:metricSetDescription:UpdateMetricSet', updateMetricSet_metricSetDescription - The dataset's description.

$sel:metricSetArn:UpdateMetricSet', updateMetricSet_metricSetArn - The ARN of the dataset to update.

Request Lenses

updateMetricSet_dimensionList :: Lens' UpdateMetricSet (Maybe (NonEmpty Text))

The dimension list.

updateMetricSet_offset :: Lens' UpdateMetricSet (Maybe Natural)

After an interval ends, the amount of seconds that the detector waits before importing data. Offset is only supported for S3 and Redshift datasources.

updateMetricSet_timestampColumn :: Lens' UpdateMetricSet (Maybe TimestampColumn)

The timestamp column.

updateMetricSet_metricList :: Lens' UpdateMetricSet (Maybe (NonEmpty Metric))

The metric list.

updateMetricSet_metricSource :: Lens' UpdateMetricSet (Maybe MetricSource)

Undocumented member.

updateMetricSet_metricSetFrequency :: Lens' UpdateMetricSet (Maybe Frequency)

The dataset's interval.

updateMetricSet_metricSetDescription :: Lens' UpdateMetricSet (Maybe Text)

The dataset's description.

updateMetricSet_metricSetArn :: Lens' UpdateMetricSet Text

The ARN of the dataset to update.

Destructuring the Response

data UpdateMetricSetResponse

See: newUpdateMetricSetResponse smart constructor.

Constructors

UpdateMetricSetResponse'

Fields

metricSetArn :: Maybe Text

The ARN of the dataset.

httpStatus :: Int

The response's http status code.

newUpdateMetricSetResponse

Arguments

:: Int
$sel:httpStatus:UpdateMetricSetResponse'

-> UpdateMetricSetResponse

Create a value of UpdateMetricSetResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:metricSetArn:UpdateMetricSet', updateMetricSetResponse_metricSetArn - The ARN of the dataset.

$sel:httpStatus:UpdateMetricSetResponse', updateMetricSetResponse_httpStatus - The response's http status code.

Response Lenses

updateMetricSetResponse_metricSetArn :: Lens' UpdateMetricSetResponse (Maybe Text)

The ARN of the dataset.

updateMetricSetResponse_httpStatus :: Lens' UpdateMetricSetResponse Int

The response's http status code.