Amazonka.LookoutVision.DeleteModel

Description

Deletes an Amazon Lookout for Vision model. You can't delete a running model. To stop a running model, use the StopModel operation.

It might take a few seconds to delete a model. To determine if a model has been deleted, call ListProjects and check if the version of the model (ModelVersion) is in the Models array.

This operation requires permissions to perform the lookoutvision:DeleteModel operation.

Creating a Request

data DeleteModel

See: newDeleteModel smart constructor.

Constructors

DeleteModel'

Fields

clientToken :: Maybe Text

ClientToken is an idempotency token that ensures a call to DeleteModel completes only once. You choose the value to pass. For example, An issue, such as an network outage, might prevent you from getting a response from DeleteModel. In this case, safely retry your call to DeleteModel by using the same ClientToken parameter value. An error occurs if the other input parameters are not the same as in the first request. Using a different value for ClientToken is considered a new call to DeleteModel. An idempotency token is active for 8 hours.

projectName :: Text

The name of the project that contains the model that you want to delete.

modelVersion :: Text

The version of the model that you want to delete.

newDeleteModel

Arguments

:: Text
$sel:projectName:DeleteModel'

-> Text
$sel:modelVersion:DeleteModel'

-> DeleteModel

Create a value of DeleteModel with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:clientToken:DeleteModel', deleteModel_clientToken - ClientToken is an idempotency token that ensures a call to DeleteModel completes only once. You choose the value to pass. For example, An issue, such as an network outage, might prevent you from getting a response from DeleteModel. In this case, safely retry your call to DeleteModel by using the same ClientToken parameter value. An error occurs if the other input parameters are not the same as in the first request. Using a different value for ClientToken is considered a new call to DeleteModel. An idempotency token is active for 8 hours.

$sel:projectName:DeleteModel', deleteModel_projectName - The name of the project that contains the model that you want to delete.

$sel:modelVersion:DeleteModel', deleteModel_modelVersion - The version of the model that you want to delete.

Request Lenses

deleteModel_clientToken :: Lens' DeleteModel (Maybe Text)

ClientToken is an idempotency token that ensures a call to DeleteModel completes only once. You choose the value to pass. For example, An issue, such as an network outage, might prevent you from getting a response from DeleteModel. In this case, safely retry your call to DeleteModel by using the same ClientToken parameter value. An error occurs if the other input parameters are not the same as in the first request. Using a different value for ClientToken is considered a new call to DeleteModel. An idempotency token is active for 8 hours.

deleteModel_projectName :: Lens' DeleteModel Text

The name of the project that contains the model that you want to delete.

deleteModel_modelVersion :: Lens' DeleteModel Text

The version of the model that you want to delete.

Destructuring the Response

data DeleteModelResponse

See: newDeleteModelResponse smart constructor.

Constructors

DeleteModelResponse'

Fields

modelArn :: Maybe Text

The Amazon Resource Name (ARN) of the model that was deleted.

httpStatus :: Int

The response's http status code.

newDeleteModelResponse

Arguments

:: Int
$sel:httpStatus:DeleteModelResponse'

-> DeleteModelResponse

Create a value of DeleteModelResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:modelArn:DeleteModelResponse', deleteModelResponse_modelArn - The Amazon Resource Name (ARN) of the model that was deleted.

$sel:httpStatus:DeleteModelResponse', deleteModelResponse_httpStatus - The response's http status code.

Response Lenses

deleteModelResponse_modelArn :: Lens' DeleteModelResponse (Maybe Text)

The Amazon Resource Name (ARN) of the model that was deleted.

deleteModelResponse_httpStatus :: Lens' DeleteModelResponse Int

The response's http status code.