Amazonka.LookoutVision.DetectAnomalies

Description

Detects anomalies in an image that you supply.

The response from DetectAnomalies includes a boolean prediction that the image contains one or more anomalies and a confidence value for the prediction.

Before calling DetectAnomalies, you must first start your model with the StartModel operation. You are charged for the amount of time, in minutes, that a model runs and for the number of anomaly detection units that your model uses. If you are not using a model, use the StopModel operation to stop your model.

This operation requires permissions to perform the lookoutvision:DetectAnomalies operation.

Creating a Request

data DetectAnomalies

See: newDetectAnomalies smart constructor.

Constructors

DetectAnomalies'

Fields

projectName :: Text

The name of the project that contains the model version that you want to use.

modelVersion :: Text

The version of the model that you want to use.

contentType :: Text

The type of the image passed in Body. Valid values are image/png (PNG format images) and image/jpeg (JPG format images).

body :: HashedBody

The unencrypted image bytes that you want to analyze.

newDetectAnomalies

Arguments

:: Text
$sel:projectName:DetectAnomalies'

-> Text
$sel:modelVersion:DetectAnomalies'

-> Text
$sel:contentType:DetectAnomalies'

-> HashedBody
$sel:body:DetectAnomalies'

-> DetectAnomalies

Create a value of DetectAnomalies with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:projectName:DetectAnomalies', detectAnomalies_projectName - The name of the project that contains the model version that you want to use.

$sel:modelVersion:DetectAnomalies', detectAnomalies_modelVersion - The version of the model that you want to use.

$sel:contentType:DetectAnomalies', detectAnomalies_contentType - The type of the image passed in Body. Valid values are image/png (PNG format images) and image/jpeg (JPG format images).

$sel:body:DetectAnomalies', detectAnomalies_body - The unencrypted image bytes that you want to analyze.

Request Lenses

detectAnomalies_projectName :: Lens' DetectAnomalies Text

The name of the project that contains the model version that you want to use.

detectAnomalies_modelVersion :: Lens' DetectAnomalies Text

The version of the model that you want to use.

detectAnomalies_contentType :: Lens' DetectAnomalies Text

The type of the image passed in Body. Valid values are image/png (PNG format images) and image/jpeg (JPG format images).

detectAnomalies_body :: Lens' DetectAnomalies HashedBody

The unencrypted image bytes that you want to analyze.

Destructuring the Response

data DetectAnomaliesResponse

See: newDetectAnomaliesResponse smart constructor.

Constructors

DetectAnomaliesResponse'

Fields

detectAnomalyResult :: Maybe DetectAnomalyResult

The results of the DetectAnomalies operation.

httpStatus :: Int

The response's http status code.

newDetectAnomaliesResponse

Arguments

:: Int
$sel:httpStatus:DetectAnomaliesResponse'

-> DetectAnomaliesResponse

Create a value of DetectAnomaliesResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:detectAnomalyResult:DetectAnomaliesResponse', detectAnomaliesResponse_detectAnomalyResult - The results of the DetectAnomalies operation.

$sel:httpStatus:DetectAnomaliesResponse', detectAnomaliesResponse_httpStatus - The response's http status code.

Response Lenses

detectAnomaliesResponse_detectAnomalyResult :: Lens' DetectAnomaliesResponse (Maybe DetectAnomalyResult)

The results of the DetectAnomalies operation.

detectAnomaliesResponse_httpStatus :: Lens' DetectAnomaliesResponse Int

The response's http status code.