Amazonka.LookoutVision.Lens

Description

Operations

ListProjects

listProjects_nextToken :: Lens' ListProjects (Maybe Text)

If the previous response was incomplete (because there is more data to retrieve), Amazon Lookout for Vision returns a pagination token in the response. You can use this pagination token to retrieve the next set of projects.

listProjects_maxResults :: Lens' ListProjects (Maybe Natural)

The maximum number of results to return per paginated call. The largest value you can specify is 100. If you specify a value greater than 100, a ValidationException error occurs. The default value is 100.

listProjectsResponse_nextToken :: Lens' ListProjectsResponse (Maybe Text)

If the response is truncated, Amazon Lookout for Vision returns this token that you can use in the subsequent request to retrieve the next set of projects.

listProjectsResponse_projects :: Lens' ListProjectsResponse (Maybe [ProjectMetadata])

A list of projects in your AWS account.

listProjectsResponse_httpStatus :: Lens' ListProjectsResponse Int

The response's http status code.

DeleteProject

deleteProject_clientToken :: Lens' DeleteProject (Maybe Text)

ClientToken is an idempotency token that ensures a call to DeleteProject completes only once. You choose the value to pass. For example, An issue, such as an network outage, might prevent you from getting a response from DeleteProject. In this case, safely retry your call to DeleteProject by using the same ClientToken parameter value. An error occurs if the other input parameters are not the same as in the first request. Using a different value for ClientToken is considered a new call to DeleteProject. An idempotency token is active for 8 hours.

deleteProject_projectName :: Lens' DeleteProject Text

The name of the project to delete.

deleteProjectResponse_projectArn :: Lens' DeleteProjectResponse (Maybe Text)

The Amazon Resource Name (ARN) of the project that was deleted.

deleteProjectResponse_httpStatus :: Lens' DeleteProjectResponse Int

The response's http status code.

DescribeDataset

describeDataset_projectName :: Lens' DescribeDataset Text

The name of the project that contains the dataset that you want to describe.

describeDataset_datasetType :: Lens' DescribeDataset Text

The type of the dataset to describe. Specify train to describe the training dataset. Specify test to describe the test dataset. If you have a single dataset project, specify train

describeDatasetResponse_datasetDescription :: Lens' DescribeDatasetResponse (Maybe DatasetDescription)

The description of the requested dataset.

describeDatasetResponse_httpStatus :: Lens' DescribeDatasetResponse Int

The response's http status code.

StopModel

stopModel_clientToken :: Lens' StopModel (Maybe Text)

ClientToken is an idempotency token that ensures a call to StopModel completes only once. You choose the value to pass. For example, An issue, such as an network outage, might prevent you from getting a response from StopModel. In this case, safely retry your call to StopModel by using the same ClientToken parameter value. An error occurs if the other input parameters are not the same as in the first request. Using a different value for ClientToken is considered a new call to StopModel. An idempotency token is active for 8 hours.

stopModel_projectName :: Lens' StopModel Text

The name of the project that contains the model that you want to stop.

stopModel_modelVersion :: Lens' StopModel Text

The version of the model that you want to stop.

stopModelResponse_status :: Lens' StopModelResponse (Maybe ModelHostingStatus)

The status of the model.

stopModelResponse_httpStatus :: Lens' StopModelResponse Int

The response's http status code.

ListDatasetEntries

listDatasetEntries_beforeCreationDate :: Lens' ListDatasetEntries (Maybe UTCTime)

Only includes entries before the specified date in the response. For example, 2020-06-23T00:00:00.

listDatasetEntries_sourceRefContains :: Lens' ListDatasetEntries (Maybe Text)

Perform a "contains" search on the values of the source-ref key within the dataset. For example a value of "IMG_17" returns all JSON Lines where the source-ref key value matches *IMG_17*.

listDatasetEntries_nextToken :: Lens' ListDatasetEntries (Maybe Text)

If the previous response was incomplete (because there is more data to retrieve), Amazon Lookout for Vision returns a pagination token in the response. You can use this pagination token to retrieve the next set of dataset entries.

listDatasetEntries_labeled :: Lens' ListDatasetEntries (Maybe Bool)

Specify true to include labeled entries, otherwise specify false. If you don't specify a value, Lookout for Vision returns all entries.

listDatasetEntries_anomalyClass :: Lens' ListDatasetEntries (Maybe Text)

Specify normal to include only normal images. Specify anomaly to only include anomalous entries. If you don't specify a value, Amazon Lookout for Vision returns normal and anomalous images.

listDatasetEntries_maxResults :: Lens' ListDatasetEntries (Maybe Natural)

The maximum number of results to return per paginated call. The largest value you can specify is 100. If you specify a value greater than 100, a ValidationException error occurs. The default value is 100.

listDatasetEntries_afterCreationDate :: Lens' ListDatasetEntries (Maybe UTCTime)

Only includes entries after the specified date in the response. For example, 2020-06-23T00:00:00.

listDatasetEntries_projectName :: Lens' ListDatasetEntries Text

The name of the project that contains the dataset that you want to list.

listDatasetEntries_datasetType :: Lens' ListDatasetEntries Text

The type of the dataset that you want to list. Specify train to list the training dataset. Specify test to list the test dataset. If you have a single dataset project, specify train.

listDatasetEntriesResponse_datasetEntries :: Lens' ListDatasetEntriesResponse (Maybe [Text])

A list of the entries (JSON Lines) within the dataset.

listDatasetEntriesResponse_nextToken :: Lens' ListDatasetEntriesResponse (Maybe Text)

If the response is truncated, Amazon Lookout for Vision returns this token that you can use in the subsequent request to retrieve the next set ofdataset entries.

listDatasetEntriesResponse_httpStatus :: Lens' ListDatasetEntriesResponse Int

The response's http status code.

ListTagsForResource

listTagsForResource_resourceArn :: Lens' ListTagsForResource Text

The Amazon Resource Name (ARN) of the model for which you want to list tags.

listTagsForResourceResponse_tags :: Lens' ListTagsForResourceResponse (Maybe [Tag])

A map of tag keys and values attached to the specified model.

listTagsForResourceResponse_httpStatus :: Lens' ListTagsForResourceResponse Int

The response's http status code.

DescribeProject

describeProject_projectName :: Lens' DescribeProject Text

The name of the project that you want to describe.

describeProjectResponse_projectDescription :: Lens' DescribeProjectResponse (Maybe ProjectDescription)

The description of the project.

describeProjectResponse_httpStatus :: Lens' DescribeProjectResponse Int

The response's http status code.

CreateModel

createModel_clientToken :: Lens' CreateModel (Maybe Text)

ClientToken is an idempotency token that ensures a call to CreateModel completes only once. You choose the value to pass. For example, An issue, such as an network outage, might prevent you from getting a response from CreateModel. In this case, safely retry your call to CreateModel by using the same ClientToken parameter value. An error occurs if the other input parameters are not the same as in the first request. Using a different value for ClientToken is considered a new call to CreateModel. An idempotency token is active for 8 hours.

createModel_kmsKeyId :: Lens' CreateModel (Maybe Text)

The identifier for your AWS Key Management Service (AWS KMS) customer master key (CMK). The key is used to encrypt training and test images copied into the service for model training. Your source images are unaffected. If this parameter is not specified, the copied images are encrypted by a key that AWS owns and manages.

createModel_description :: Lens' CreateModel (Maybe Text)

A description for the version of the model.

createModel_tags :: Lens' CreateModel (Maybe [Tag])

A set of tags (key-value pairs) that you want to attach to the model.

createModel_projectName :: Lens' CreateModel Text

The name of the project in which you want to create a model version.

createModel_outputConfig :: Lens' CreateModel OutputConfig

The location where Amazon Lookout for Vision saves the training results.

createModelResponse_modelMetadata :: Lens' CreateModelResponse (Maybe ModelMetadata)

The response from a call to CreateModel.

createModelResponse_httpStatus :: Lens' CreateModelResponse Int

The response's http status code.

DeleteDataset

deleteDataset_clientToken :: Lens' DeleteDataset (Maybe Text)

ClientToken is an idempotency token that ensures a call to DeleteDataset completes only once. You choose the value to pass. For example, An issue, such as an network outage, might prevent you from getting a response from DeleteDataset. In this case, safely retry your call to DeleteDataset by using the same ClientToken parameter value. An error occurs if the other input parameters are not the same as in the first request. Using a different value for ClientToken is considered a new call to DeleteDataset. An idempotency token is active for 8 hours.

deleteDataset_projectName :: Lens' DeleteDataset Text

The name of the project that contains the dataset that you want to delete.

deleteDataset_datasetType :: Lens' DeleteDataset Text

The type of the dataset to delete. Specify train to delete the training dataset. Specify test to delete the test dataset. To delete the dataset in a single dataset project, specify train.

deleteDatasetResponse_httpStatus :: Lens' DeleteDatasetResponse Int

The response's http status code.

CreateDataset

createDataset_clientToken :: Lens' CreateDataset (Maybe Text)

ClientToken is an idempotency token that ensures a call to CreateDataset completes only once. You choose the value to pass. For example, An issue, such as an network outage, might prevent you from getting a response from CreateDataset. In this case, safely retry your call to CreateDataset by using the same ClientToken parameter value. An error occurs if the other input parameters are not the same as in the first request. Using a different value for ClientToken is considered a new call to CreateDataset. An idempotency token is active for 8 hours.

createDataset_datasetSource :: Lens' CreateDataset (Maybe DatasetSource)

The location of the manifest file that Amazon Lookout for Vision uses to create the dataset.

If you don't specify DatasetSource, an empty dataset is created and the operation synchronously returns. Later, you can add JSON Lines by calling UpdateDatasetEntries.

If you specify a value for DataSource, the manifest at the S3 location is validated and used to create the dataset. The call to CreateDataset is asynchronous and might take a while to complete. To find out the current status, Check the value of Status returned in a call to DescribeDataset.

createDataset_projectName :: Lens' CreateDataset Text

The name of the project in which you want to create a dataset.

createDataset_datasetType :: Lens' CreateDataset Text

The type of the dataset. Specify train for a training dataset. Specify test for a test dataset.

createDatasetResponse_datasetMetadata :: Lens' CreateDatasetResponse (Maybe DatasetMetadata)

Information about the dataset.

createDatasetResponse_httpStatus :: Lens' CreateDatasetResponse Int

The response's http status code.

DeleteModel

deleteModel_clientToken :: Lens' DeleteModel (Maybe Text)

ClientToken is an idempotency token that ensures a call to DeleteModel completes only once. You choose the value to pass. For example, An issue, such as an network outage, might prevent you from getting a response from DeleteModel. In this case, safely retry your call to DeleteModel by using the same ClientToken parameter value. An error occurs if the other input parameters are not the same as in the first request. Using a different value for ClientToken is considered a new call to DeleteModel. An idempotency token is active for 8 hours.

deleteModel_projectName :: Lens' DeleteModel Text

The name of the project that contains the model that you want to delete.

deleteModel_modelVersion :: Lens' DeleteModel Text

The version of the model that you want to delete.

deleteModelResponse_modelArn :: Lens' DeleteModelResponse (Maybe Text)

The Amazon Resource Name (ARN) of the model that was deleted.

deleteModelResponse_httpStatus :: Lens' DeleteModelResponse Int

The response's http status code.

ListModels

listModels_nextToken :: Lens' ListModels (Maybe Text)

If the previous response was incomplete (because there is more data to retrieve), Amazon Lookout for Vision returns a pagination token in the response. You can use this pagination token to retrieve the next set of models.

listModels_maxResults :: Lens' ListModels (Maybe Natural)

The maximum number of results to return per paginated call. The largest value you can specify is 100. If you specify a value greater than 100, a ValidationException error occurs. The default value is 100.

listModels_projectName :: Lens' ListModels Text

The name of the project that contains the model versions that you want to list.

listModelsResponse_nextToken :: Lens' ListModelsResponse (Maybe Text)

If the response is truncated, Amazon Lookout for Vision returns this token that you can use in the subsequent request to retrieve the next set of models.

listModelsResponse_models :: Lens' ListModelsResponse (Maybe [ModelMetadata])

A list of model versions in the specified project.

listModelsResponse_httpStatus :: Lens' ListModelsResponse Int

The response's http status code.

StartModel

startModel_clientToken :: Lens' StartModel (Maybe Text)

ClientToken is an idempotency token that ensures a call to StartModel completes only once. You choose the value to pass. For example, An issue, such as an network outage, might prevent you from getting a response from StartModel. In this case, safely retry your call to StartModel by using the same ClientToken parameter value. An error occurs if the other input parameters are not the same as in the first request. Using a different value for ClientToken is considered a new call to StartModel. An idempotency token is active for 8 hours.

startModel_projectName :: Lens' StartModel Text

The name of the project that contains the model that you want to start.

startModel_modelVersion :: Lens' StartModel Text

The version of the model that you want to start.

startModel_minInferenceUnits :: Lens' StartModel Natural

The minimum number of inference units to use. A single inference unit represents 1 hour of processing and can support up to 5 Transaction Pers Second (TPS). Use a higher number to increase the TPS throughput of your model. You are charged for the number of inference units that you use.

startModelResponse_status :: Lens' StartModelResponse (Maybe ModelHostingStatus)

The current running status of the model.

startModelResponse_httpStatus :: Lens' StartModelResponse Int

The response's http status code.

DescribeModel

describeModel_projectName :: Lens' DescribeModel Text

The project that contains the version of a model that you want to describe.

describeModel_modelVersion :: Lens' DescribeModel Text

The version of the model that you want to describe.

describeModelResponse_modelDescription :: Lens' DescribeModelResponse (Maybe ModelDescription)

Contains the description of the model.

describeModelResponse_httpStatus :: Lens' DescribeModelResponse Int

The response's http status code.

DetectAnomalies

detectAnomalies_projectName :: Lens' DetectAnomalies Text

The name of the project that contains the model version that you want to use.

detectAnomalies_modelVersion :: Lens' DetectAnomalies Text

The version of the model that you want to use.

detectAnomalies_contentType :: Lens' DetectAnomalies Text

The type of the image passed in Body. Valid values are image/png (PNG format images) and image/jpeg (JPG format images).

detectAnomalies_body :: Lens' DetectAnomalies HashedBody

The unencrypted image bytes that you want to analyze.

detectAnomaliesResponse_detectAnomalyResult :: Lens' DetectAnomaliesResponse (Maybe DetectAnomalyResult)

The results of the DetectAnomalies operation.

detectAnomaliesResponse_httpStatus :: Lens' DetectAnomaliesResponse Int

The response's http status code.

TagResource

tagResource_resourceArn :: Lens' TagResource Text

The Amazon Resource Name (ARN) of the model to assign the tags.

tagResource_tags :: Lens' TagResource [Tag]

The key-value tags to assign to the model.

tagResourceResponse_httpStatus :: Lens' TagResourceResponse Int

The response's http status code.

UntagResource

untagResource_resourceArn :: Lens' UntagResource Text

The Amazon Resource Name (ARN) of the model from which you want to remove tags.

untagResource_tagKeys :: Lens' UntagResource [Text]

A list of the keys of the tags that you want to remove.

untagResourceResponse_httpStatus :: Lens' UntagResourceResponse Int

The response's http status code.

CreateProject

createProject_clientToken :: Lens' CreateProject (Maybe Text)

ClientToken is an idempotency token that ensures a call to CreateProject completes only once. You choose the value to pass. For example, An issue, such as an network outage, might prevent you from getting a response from CreateProject. In this case, safely retry your call to CreateProject by using the same ClientToken parameter value. An error occurs if the other input parameters are not the same as in the first request. Using a different value for ClientToken is considered a new call to CreateProject. An idempotency token is active for 8 hours.

createProject_projectName :: Lens' CreateProject Text

The name for the project.

createProjectResponse_projectMetadata :: Lens' CreateProjectResponse (Maybe ProjectMetadata)

Information about the project.

createProjectResponse_httpStatus :: Lens' CreateProjectResponse Int

The response's http status code.

UpdateDatasetEntries

updateDatasetEntries_clientToken :: Lens' UpdateDatasetEntries (Maybe Text)

ClientToken is an idempotency token that ensures a call to UpdateDatasetEntries completes only once. You choose the value to pass. For example, An issue, such as an network outage, might prevent you from getting a response from UpdateDatasetEntries. In this case, safely retry your call to UpdateDatasetEntries by using the same ClientToken parameter value. An error occurs if the other input parameters are not the same as in the first request. Using a different value for ClientToken is considered a new call to UpdateDatasetEntries. An idempotency token is active for 8 hours.

updateDatasetEntries_projectName :: Lens' UpdateDatasetEntries Text

The name of the project that contains the dataset that you want to update.

updateDatasetEntries_datasetType :: Lens' UpdateDatasetEntries Text

The type of the dataset that you want to update. Specify train to update the training dataset. Specify test to update the test dataset. If you have a single dataset project, specify train.

updateDatasetEntries_changes :: Lens' UpdateDatasetEntries ByteString

The entries to add to the dataset.-- -- Note: This Lens automatically encodes and decodes Base64 data. -- The underlying isomorphism will encode to Base64 representation during -- serialisation, and decode from Base64 representation during deserialisation. -- This Lens accepts and returns only raw unencoded data.

updateDatasetEntriesResponse_status :: Lens' UpdateDatasetEntriesResponse (Maybe DatasetStatus)

The status of the dataset update.

updateDatasetEntriesResponse_httpStatus :: Lens' UpdateDatasetEntriesResponse Int

The response's http status code.

Types

DatasetDescription

datasetDescription_status :: Lens' DatasetDescription (Maybe DatasetStatus)

The status of the dataset.

datasetDescription_imageStats :: Lens' DatasetDescription (Maybe DatasetImageStats)

datasetDescription_statusMessage :: Lens' DatasetDescription (Maybe Text)

The status message for the dataset.

datasetDescription_creationTimestamp :: Lens' DatasetDescription (Maybe UTCTime)

The Unix timestamp for the time and date that the dataset was created.

datasetDescription_datasetType :: Lens' DatasetDescription (Maybe Text)

The type of the dataset. The value train represents a training dataset or single dataset project. The value test represents a test dataset.

datasetDescription_projectName :: Lens' DatasetDescription (Maybe Text)

The name of the project that contains the dataset.

datasetDescription_lastUpdatedTimestamp :: Lens' DatasetDescription (Maybe UTCTime)

The Unix timestamp for the date and time that the dataset was last updated.

DatasetGroundTruthManifest

datasetGroundTruthManifest_s3Object :: Lens' DatasetGroundTruthManifest (Maybe InputS3Object)

The S3 bucket location for the manifest file.

DatasetImageStats

datasetImageStats_normal :: Lens' DatasetImageStats (Maybe Int)

The total number of images labeled as normal.

datasetImageStats_anomaly :: Lens' DatasetImageStats (Maybe Int)

the total number of images labeled as an anomaly.

datasetImageStats_labeled :: Lens' DatasetImageStats (Maybe Int)

The total number of labeled images.

datasetImageStats_total :: Lens' DatasetImageStats (Maybe Int)

The total number of images in the dataset.

DatasetMetadata

datasetMetadata_status :: Lens' DatasetMetadata (Maybe DatasetStatus)

The status for the dataset.

datasetMetadata_statusMessage :: Lens' DatasetMetadata (Maybe Text)

The status message for the dataset.

datasetMetadata_creationTimestamp :: Lens' DatasetMetadata (Maybe UTCTime)

The Unix timestamp for the date and time that the dataset was created.

datasetMetadata_datasetType :: Lens' DatasetMetadata (Maybe Text)

The type of the dataset.

DatasetSource

datasetSource_groundTruthManifest :: Lens' DatasetSource (Maybe DatasetGroundTruthManifest)

Location information for the manifest file.

DetectAnomalyResult

detectAnomalyResult_isAnomalous :: Lens' DetectAnomalyResult (Maybe Bool)

True if the image contains an anomaly, otherwise false.

detectAnomalyResult_confidence :: Lens' DetectAnomalyResult (Maybe Double)

The confidence that Amazon Lookout for Vision has in the accuracy of the prediction.

detectAnomalyResult_source :: Lens' DetectAnomalyResult (Maybe ImageSource)

The source of the image that was analyzed. direct means that the images was supplied from the local computer. No other values are supported.

ImageSource

imageSource_type :: Lens' ImageSource (Maybe Text)

The type of the image.

InputS3Object

inputS3Object_versionId :: Lens' InputS3Object (Maybe Text)

The version ID of the bucket.

inputS3Object_bucket :: Lens' InputS3Object Text

The Amazon S3 bucket that contains the manifest.

inputS3Object_key :: Lens' InputS3Object Text

The name and location of the manifest file withiin the bucket.

ModelDescription

modelDescription_status :: Lens' ModelDescription (Maybe ModelStatus)

The status of the model.

modelDescription_evaluationResult :: Lens' ModelDescription (Maybe OutputS3Object)

The S3 location where Amazon Lookout for Vision saves the performance metrics.

modelDescription_evaluationEndTimestamp :: Lens' ModelDescription (Maybe UTCTime)

The unix timestamp for the date and time that the evaluation ended.

modelDescription_modelArn :: Lens' ModelDescription (Maybe Text)

The Amazon Resource Name (ARN) of the model.

modelDescription_performance :: Lens' ModelDescription (Maybe ModelPerformance)

Performance metrics for the model. Created during training.

modelDescription_kmsKeyId :: Lens' ModelDescription (Maybe Text)

The identifer for the AWS Key Management Service (AWS KMS) key that was used to encrypt the model during training.

modelDescription_statusMessage :: Lens' ModelDescription (Maybe Text)

The status message for the model.

modelDescription_creationTimestamp :: Lens' ModelDescription (Maybe UTCTime)

The unix timestamp for the date and time that the model was created.

modelDescription_outputConfig :: Lens' ModelDescription (Maybe OutputConfig)

The S3 location where Amazon Lookout for Vision saves model training files.

modelDescription_modelVersion :: Lens' ModelDescription (Maybe Text)

The version of the model

modelDescription_description :: Lens' ModelDescription (Maybe Text)

The description for the model.

modelDescription_evaluationManifest :: Lens' ModelDescription (Maybe OutputS3Object)

The S3 location where Amazon Lookout for Vision saves the manifest file that was used to test the trained model and generate the performance scores.

ModelMetadata

modelMetadata_status :: Lens' ModelMetadata (Maybe ModelStatus)

The status of the model.

modelMetadata_modelArn :: Lens' ModelMetadata (Maybe Text)

The Amazon Resource Name (ARN) of the model.

modelMetadata_performance :: Lens' ModelMetadata (Maybe ModelPerformance)

Performance metrics for the model. Not available until training has successfully completed.

modelMetadata_statusMessage :: Lens' ModelMetadata (Maybe Text)

The status message for the model.

modelMetadata_creationTimestamp :: Lens' ModelMetadata (Maybe UTCTime)

The unix timestamp for the date and time that the model was created.

modelMetadata_modelVersion :: Lens' ModelMetadata (Maybe Text)

The version of the model.

modelMetadata_description :: Lens' ModelMetadata (Maybe Text)

The description for the model.

ModelPerformance

modelPerformance_recall :: Lens' ModelPerformance (Maybe Double)

The overall recall metric value for the trained model.

modelPerformance_precision :: Lens' ModelPerformance (Maybe Double)

The overall precision metric value for the trained model.

modelPerformance_f1Score :: Lens' ModelPerformance (Maybe Double)

The overall F1 score metric for the trained model.

OutputConfig

outputConfig_s3Location :: Lens' OutputConfig S3Location

The S3 location for the output.

OutputS3Object

outputS3Object_bucket :: Lens' OutputS3Object Text

The bucket that contains the training output.

outputS3Object_key :: Lens' OutputS3Object Text

The location of the training output in the bucket.

ProjectDescription

projectDescription_creationTimestamp :: Lens' ProjectDescription (Maybe UTCTime)

The unix timestamp for the date and time that the project was created.

projectDescription_projectName :: Lens' ProjectDescription (Maybe Text)

The name of the project.

projectDescription_projectArn :: Lens' ProjectDescription (Maybe Text)

The Amazon Resource Name (ARN) of the project.

projectDescription_datasets :: Lens' ProjectDescription (Maybe [DatasetMetadata])

A list of datasets in the project.

ProjectMetadata

projectMetadata_creationTimestamp :: Lens' ProjectMetadata (Maybe UTCTime)

The unix timestamp for the date and time that the project was created.

projectMetadata_projectName :: Lens' ProjectMetadata (Maybe Text)

The name of the project.

projectMetadata_projectArn :: Lens' ProjectMetadata (Maybe Text)

The Amazon Resource Name (ARN) of the project.

S3Location

s3Location_prefix :: Lens' S3Location (Maybe Text)

The path of the folder, within the S3 bucket, that contains the training output.

s3Location_bucket :: Lens' S3Location Text

The S3 bucket that contains the training output.

Tag

tag_key :: Lens' Tag Text

The key of the tag that is attached to the specified model.

tag_value :: Lens' Tag Text

The value of the tag that is attached to the specified model.