Amazonka.LookoutVision.Types

Description

Service Configuration

defaultService :: Service

API version 2020-11-20 of the Amazon Lookout for Vision SDK configuration.

Errors

_ValidationException :: AsError a => Getting (First ServiceError) a ServiceError

An input validation error occured. For example, invalid characters in a project name, or if a pagination token is invalid.

_AccessDeniedException :: AsError a => Getting (First ServiceError) a ServiceError

You are not authorized to perform the action.

_ConflictException :: AsError a => Getting (First ServiceError) a ServiceError

The update or deletion of a resource caused an inconsistent state.

_ServiceQuotaExceededException :: AsError a => Getting (First ServiceError) a ServiceError

A service quota was exceeded the allowed limit. For more information, see Limits in Amazon Lookout for Vision in the Amazon Lookout for Vision Developer Guide.

_ThrottlingException :: AsError a => Getting (First ServiceError) a ServiceError

Amazon Lookout for Vision is temporarily unable to process the request. Try your call again.

_InternalServerException :: AsError a => Getting (First ServiceError) a ServiceError

Amazon Lookout for Vision experienced a service issue. Try your call again.

_ResourceNotFoundException :: AsError a => Getting (First ServiceError) a ServiceError

The resource could not be found.

DatasetStatus

newtype DatasetStatus

Constructors

DatasetStatus'

Fields

fromDatasetStatus :: Text

Bundled Patterns

pattern DatasetStatus_CREATE_COMPLETE :: DatasetStatus

pattern DatasetStatus_CREATE_FAILED :: DatasetStatus

pattern DatasetStatus_CREATE_IN_PROGRESS :: DatasetStatus

pattern DatasetStatus_DELETE_COMPLETE :: DatasetStatus

pattern DatasetStatus_DELETE_FAILED :: DatasetStatus

pattern DatasetStatus_DELETE_IN_PROGRESS :: DatasetStatus

pattern DatasetStatus_UPDATE_COMPLETE :: DatasetStatus

pattern DatasetStatus_UPDATE_FAILED_ROLLBACK_COMPLETE :: DatasetStatus

pattern DatasetStatus_UPDATE_FAILED_ROLLBACK_IN_PROGRESS :: DatasetStatus

pattern DatasetStatus_UPDATE_IN_PROGRESS :: DatasetStatus

ModelHostingStatus

newtype ModelHostingStatus

Constructors

ModelHostingStatus'

Fields

fromModelHostingStatus :: Text

Bundled Patterns

pattern ModelHostingStatus_HOSTED :: ModelHostingStatus

pattern ModelHostingStatus_HOSTING_FAILED :: ModelHostingStatus

pattern ModelHostingStatus_STARTING_HOSTING :: ModelHostingStatus

pattern ModelHostingStatus_STOPPING_HOSTING :: ModelHostingStatus

pattern ModelHostingStatus_SYSTEM_UPDATING :: ModelHostingStatus

ModelStatus

newtype ModelStatus

Constructors

ModelStatus'

Fields

fromModelStatus :: Text

Bundled Patterns

pattern ModelStatus_DELETING :: ModelStatus

pattern ModelStatus_HOSTED :: ModelStatus

pattern ModelStatus_HOSTING_FAILED :: ModelStatus

pattern ModelStatus_STARTING_HOSTING :: ModelStatus

pattern ModelStatus_STOPPING_HOSTING :: ModelStatus

pattern ModelStatus_SYSTEM_UPDATING :: ModelStatus

pattern ModelStatus_TRAINED :: ModelStatus

pattern ModelStatus_TRAINING :: ModelStatus

pattern ModelStatus_TRAINING_FAILED :: ModelStatus

DatasetDescription

data DatasetDescription

The description for a dataset. For more information, see DescribeDataset.

See: newDatasetDescription smart constructor.

Constructors

DatasetDescription'

Fields

status :: Maybe DatasetStatus

The status of the dataset.

imageStats :: Maybe DatasetImageStats

statusMessage :: Maybe Text

The status message for the dataset.

creationTimestamp :: Maybe POSIX

The Unix timestamp for the time and date that the dataset was created.

datasetType :: Maybe Text

The type of the dataset. The value train represents a training dataset or single dataset project. The value test represents a test dataset.

projectName :: Maybe Text

The name of the project that contains the dataset.

lastUpdatedTimestamp :: Maybe POSIX

The Unix timestamp for the date and time that the dataset was last updated.

newDatasetDescription :: DatasetDescription

Create a value of DatasetDescription with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:status:DatasetDescription', datasetDescription_status - The status of the dataset.

$sel:imageStats:DatasetDescription', datasetDescription_imageStats -

$sel:statusMessage:DatasetDescription', datasetDescription_statusMessage - The status message for the dataset.

$sel:creationTimestamp:DatasetDescription', datasetDescription_creationTimestamp - The Unix timestamp for the time and date that the dataset was created.

$sel:datasetType:DatasetDescription', datasetDescription_datasetType - The type of the dataset. The value train represents a training dataset or single dataset project. The value test represents a test dataset.

$sel:projectName:DatasetDescription', datasetDescription_projectName - The name of the project that contains the dataset.

$sel:lastUpdatedTimestamp:DatasetDescription', datasetDescription_lastUpdatedTimestamp - The Unix timestamp for the date and time that the dataset was last updated.

datasetDescription_status :: Lens' DatasetDescription (Maybe DatasetStatus)

The status of the dataset.

datasetDescription_imageStats :: Lens' DatasetDescription (Maybe DatasetImageStats)

datasetDescription_statusMessage :: Lens' DatasetDescription (Maybe Text)

The status message for the dataset.

datasetDescription_creationTimestamp :: Lens' DatasetDescription (Maybe UTCTime)

The Unix timestamp for the time and date that the dataset was created.

datasetDescription_datasetType :: Lens' DatasetDescription (Maybe Text)

The type of the dataset. The value train represents a training dataset or single dataset project. The value test represents a test dataset.

datasetDescription_projectName :: Lens' DatasetDescription (Maybe Text)

The name of the project that contains the dataset.

datasetDescription_lastUpdatedTimestamp :: Lens' DatasetDescription (Maybe UTCTime)

The Unix timestamp for the date and time that the dataset was last updated.

DatasetGroundTruthManifest

data DatasetGroundTruthManifest

Location information about a manifest file. You can use a manifest file to create a dataset.

See: newDatasetGroundTruthManifest smart constructor.

Constructors

DatasetGroundTruthManifest'

Fields

s3Object :: Maybe InputS3Object

The S3 bucket location for the manifest file.

newDatasetGroundTruthManifest :: DatasetGroundTruthManifest

Create a value of DatasetGroundTruthManifest with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:s3Object:DatasetGroundTruthManifest', datasetGroundTruthManifest_s3Object - The S3 bucket location for the manifest file.

datasetGroundTruthManifest_s3Object :: Lens' DatasetGroundTruthManifest (Maybe InputS3Object)

The S3 bucket location for the manifest file.

DatasetImageStats

data DatasetImageStats

Statistics about the images in a dataset.

See: newDatasetImageStats smart constructor.

Constructors

DatasetImageStats'

Fields

normal :: Maybe Int

The total number of images labeled as normal.

anomaly :: Maybe Int

the total number of images labeled as an anomaly.

labeled :: Maybe Int

The total number of labeled images.

total :: Maybe Int

The total number of images in the dataset.

newDatasetImageStats :: DatasetImageStats

Create a value of DatasetImageStats with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:normal:DatasetImageStats', datasetImageStats_normal - The total number of images labeled as normal.

$sel:anomaly:DatasetImageStats', datasetImageStats_anomaly - the total number of images labeled as an anomaly.

$sel:labeled:DatasetImageStats', datasetImageStats_labeled - The total number of labeled images.

$sel:total:DatasetImageStats', datasetImageStats_total - The total number of images in the dataset.

datasetImageStats_normal :: Lens' DatasetImageStats (Maybe Int)

The total number of images labeled as normal.

datasetImageStats_anomaly :: Lens' DatasetImageStats (Maybe Int)

the total number of images labeled as an anomaly.

datasetImageStats_labeled :: Lens' DatasetImageStats (Maybe Int)

The total number of labeled images.

datasetImageStats_total :: Lens' DatasetImageStats (Maybe Int)

The total number of images in the dataset.

DatasetMetadata

data DatasetMetadata

Sumary information for an Amazon Lookout for Vision dataset.

See: newDatasetMetadata smart constructor.

Constructors

DatasetMetadata'

Fields

status :: Maybe DatasetStatus

The status for the dataset.

statusMessage :: Maybe Text

The status message for the dataset.

creationTimestamp :: Maybe POSIX

The Unix timestamp for the date and time that the dataset was created.

datasetType :: Maybe Text

The type of the dataset.

newDatasetMetadata :: DatasetMetadata

Create a value of DatasetMetadata with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:status:DatasetMetadata', datasetMetadata_status - The status for the dataset.

$sel:statusMessage:DatasetMetadata', datasetMetadata_statusMessage - The status message for the dataset.

$sel:creationTimestamp:DatasetMetadata', datasetMetadata_creationTimestamp - The Unix timestamp for the date and time that the dataset was created.

$sel:datasetType:DatasetMetadata', datasetMetadata_datasetType - The type of the dataset.

datasetMetadata_status :: Lens' DatasetMetadata (Maybe DatasetStatus)

The status for the dataset.

datasetMetadata_statusMessage :: Lens' DatasetMetadata (Maybe Text)

The status message for the dataset.

datasetMetadata_creationTimestamp :: Lens' DatasetMetadata (Maybe UTCTime)

The Unix timestamp for the date and time that the dataset was created.

datasetMetadata_datasetType :: Lens' DatasetMetadata (Maybe Text)

The type of the dataset.

DatasetSource

data DatasetSource

Information about the location of a manifest file that Amazon Lookout for Vision uses to to create a dataset.

See: newDatasetSource smart constructor.

Constructors

DatasetSource'

Fields

groundTruthManifest :: Maybe DatasetGroundTruthManifest

Location information for the manifest file.

newDatasetSource :: DatasetSource

Create a value of DatasetSource with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:groundTruthManifest:DatasetSource', datasetSource_groundTruthManifest - Location information for the manifest file.

datasetSource_groundTruthManifest :: Lens' DatasetSource (Maybe DatasetGroundTruthManifest)

Location information for the manifest file.

DetectAnomalyResult

data DetectAnomalyResult

The prediction results from a call to DetectAnomalies.

See: newDetectAnomalyResult smart constructor.

Constructors

DetectAnomalyResult'

Fields

isAnomalous :: Maybe Bool

True if the image contains an anomaly, otherwise false.

confidence :: Maybe Double

The confidence that Amazon Lookout for Vision has in the accuracy of the prediction.

source :: Maybe ImageSource

The source of the image that was analyzed. direct means that the images was supplied from the local computer. No other values are supported.

newDetectAnomalyResult :: DetectAnomalyResult

Create a value of DetectAnomalyResult with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:isAnomalous:DetectAnomalyResult', detectAnomalyResult_isAnomalous - True if the image contains an anomaly, otherwise false.

$sel:confidence:DetectAnomalyResult', detectAnomalyResult_confidence - The confidence that Amazon Lookout for Vision has in the accuracy of the prediction.

$sel:source:DetectAnomalyResult', detectAnomalyResult_source - The source of the image that was analyzed. direct means that the images was supplied from the local computer. No other values are supported.

detectAnomalyResult_isAnomalous :: Lens' DetectAnomalyResult (Maybe Bool)

True if the image contains an anomaly, otherwise false.

detectAnomalyResult_confidence :: Lens' DetectAnomalyResult (Maybe Double)

The confidence that Amazon Lookout for Vision has in the accuracy of the prediction.

detectAnomalyResult_source :: Lens' DetectAnomalyResult (Maybe ImageSource)

The source of the image that was analyzed. direct means that the images was supplied from the local computer. No other values are supported.

ImageSource

data ImageSource

The source for an image.

See: newImageSource smart constructor.

Constructors

ImageSource'

Fields

type' :: Maybe Text

The type of the image.

newImageSource :: ImageSource

Create a value of ImageSource with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:type':ImageSource', imageSource_type - The type of the image.

imageSource_type :: Lens' ImageSource (Maybe Text)

The type of the image.

InputS3Object

data InputS3Object

Amazon S3 Location information for an input manifest file.

See: newInputS3Object smart constructor.

Constructors

InputS3Object'

Fields

versionId :: Maybe Text

The version ID of the bucket.

bucket :: Text

The Amazon S3 bucket that contains the manifest.

key :: Text

The name and location of the manifest file withiin the bucket.

newInputS3Object

Arguments

:: Text
$sel:bucket:InputS3Object'

-> Text
$sel:key:InputS3Object'

-> InputS3Object

Create a value of InputS3Object with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:versionId:InputS3Object', inputS3Object_versionId - The version ID of the bucket.

$sel:bucket:InputS3Object', inputS3Object_bucket - The Amazon S3 bucket that contains the manifest.

$sel:key:InputS3Object', inputS3Object_key - The name and location of the manifest file withiin the bucket.

inputS3Object_versionId :: Lens' InputS3Object (Maybe Text)

The version ID of the bucket.

inputS3Object_bucket :: Lens' InputS3Object Text

The Amazon S3 bucket that contains the manifest.

inputS3Object_key :: Lens' InputS3Object Text

The name and location of the manifest file withiin the bucket.

ModelDescription

data ModelDescription

Describes an Amazon Lookout for Vision model.

See: newModelDescription smart constructor.

Constructors

ModelDescription'

Fields

status :: Maybe ModelStatus

The status of the model.

evaluationResult :: Maybe OutputS3Object

The S3 location where Amazon Lookout for Vision saves the performance metrics.

evaluationEndTimestamp :: Maybe POSIX

The unix timestamp for the date and time that the evaluation ended.

modelArn :: Maybe Text

The Amazon Resource Name (ARN) of the model.

performance :: Maybe ModelPerformance

Performance metrics for the model. Created during training.

kmsKeyId :: Maybe Text

The identifer for the AWS Key Management Service (AWS KMS) key that was used to encrypt the model during training.

statusMessage :: Maybe Text

The status message for the model.

creationTimestamp :: Maybe POSIX

The unix timestamp for the date and time that the model was created.

outputConfig :: Maybe OutputConfig

The S3 location where Amazon Lookout for Vision saves model training files.

modelVersion :: Maybe Text

The version of the model

description :: Maybe Text

The description for the model.

evaluationManifest :: Maybe OutputS3Object

The S3 location where Amazon Lookout for Vision saves the manifest file that was used to test the trained model and generate the performance scores.

newModelDescription :: ModelDescription

Create a value of ModelDescription with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:status:ModelDescription', modelDescription_status - The status of the model.

$sel:evaluationResult:ModelDescription', modelDescription_evaluationResult - The S3 location where Amazon Lookout for Vision saves the performance metrics.

$sel:evaluationEndTimestamp:ModelDescription', modelDescription_evaluationEndTimestamp - The unix timestamp for the date and time that the evaluation ended.

$sel:modelArn:ModelDescription', modelDescription_modelArn - The Amazon Resource Name (ARN) of the model.

$sel:performance:ModelDescription', modelDescription_performance - Performance metrics for the model. Created during training.

$sel:kmsKeyId:ModelDescription', modelDescription_kmsKeyId - The identifer for the AWS Key Management Service (AWS KMS) key that was used to encrypt the model during training.

$sel:statusMessage:ModelDescription', modelDescription_statusMessage - The status message for the model.

$sel:creationTimestamp:ModelDescription', modelDescription_creationTimestamp - The unix timestamp for the date and time that the model was created.

$sel:outputConfig:ModelDescription', modelDescription_outputConfig - The S3 location where Amazon Lookout for Vision saves model training files.

$sel:modelVersion:ModelDescription', modelDescription_modelVersion - The version of the model

$sel:description:ModelDescription', modelDescription_description - The description for the model.

$sel:evaluationManifest:ModelDescription', modelDescription_evaluationManifest - The S3 location where Amazon Lookout for Vision saves the manifest file that was used to test the trained model and generate the performance scores.

modelDescription_status :: Lens' ModelDescription (Maybe ModelStatus)

The status of the model.

modelDescription_evaluationResult :: Lens' ModelDescription (Maybe OutputS3Object)

The S3 location where Amazon Lookout for Vision saves the performance metrics.

modelDescription_evaluationEndTimestamp :: Lens' ModelDescription (Maybe UTCTime)

The unix timestamp for the date and time that the evaluation ended.

modelDescription_modelArn :: Lens' ModelDescription (Maybe Text)

The Amazon Resource Name (ARN) of the model.

modelDescription_performance :: Lens' ModelDescription (Maybe ModelPerformance)

Performance metrics for the model. Created during training.

modelDescription_kmsKeyId :: Lens' ModelDescription (Maybe Text)

The identifer for the AWS Key Management Service (AWS KMS) key that was used to encrypt the model during training.

modelDescription_statusMessage :: Lens' ModelDescription (Maybe Text)

The status message for the model.

modelDescription_creationTimestamp :: Lens' ModelDescription (Maybe UTCTime)

The unix timestamp for the date and time that the model was created.

modelDescription_outputConfig :: Lens' ModelDescription (Maybe OutputConfig)

The S3 location where Amazon Lookout for Vision saves model training files.

modelDescription_modelVersion :: Lens' ModelDescription (Maybe Text)

The version of the model

modelDescription_description :: Lens' ModelDescription (Maybe Text)

The description for the model.

modelDescription_evaluationManifest :: Lens' ModelDescription (Maybe OutputS3Object)

The S3 location where Amazon Lookout for Vision saves the manifest file that was used to test the trained model and generate the performance scores.

ModelMetadata

data ModelMetadata

Describes an Amazon Lookout for Vision model.

See: newModelMetadata smart constructor.

Constructors

ModelMetadata'

Fields

status :: Maybe ModelStatus

The status of the model.

modelArn :: Maybe Text

The Amazon Resource Name (ARN) of the model.

performance :: Maybe ModelPerformance

Performance metrics for the model. Not available until training has successfully completed.

statusMessage :: Maybe Text

The status message for the model.

creationTimestamp :: Maybe POSIX

The unix timestamp for the date and time that the model was created.

modelVersion :: Maybe Text

The version of the model.

description :: Maybe Text

The description for the model.

newModelMetadata :: ModelMetadata

Create a value of ModelMetadata with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:status:ModelMetadata', modelMetadata_status - The status of the model.

$sel:modelArn:ModelMetadata', modelMetadata_modelArn - The Amazon Resource Name (ARN) of the model.

$sel:performance:ModelMetadata', modelMetadata_performance - Performance metrics for the model. Not available until training has successfully completed.

$sel:statusMessage:ModelMetadata', modelMetadata_statusMessage - The status message for the model.

$sel:creationTimestamp:ModelMetadata', modelMetadata_creationTimestamp - The unix timestamp for the date and time that the model was created.

$sel:modelVersion:ModelMetadata', modelMetadata_modelVersion - The version of the model.

$sel:description:ModelMetadata', modelMetadata_description - The description for the model.

modelMetadata_status :: Lens' ModelMetadata (Maybe ModelStatus)

The status of the model.

modelMetadata_modelArn :: Lens' ModelMetadata (Maybe Text)

The Amazon Resource Name (ARN) of the model.

modelMetadata_performance :: Lens' ModelMetadata (Maybe ModelPerformance)

Performance metrics for the model. Not available until training has successfully completed.

modelMetadata_statusMessage :: Lens' ModelMetadata (Maybe Text)

The status message for the model.

modelMetadata_creationTimestamp :: Lens' ModelMetadata (Maybe UTCTime)

The unix timestamp for the date and time that the model was created.

modelMetadata_modelVersion :: Lens' ModelMetadata (Maybe Text)

The version of the model.

modelMetadata_description :: Lens' ModelMetadata (Maybe Text)

The description for the model.

ModelPerformance

data ModelPerformance

Information about the evaluation performance of a trained model.

See: newModelPerformance smart constructor.

Constructors

ModelPerformance'

Fields

recall :: Maybe Double

The overall recall metric value for the trained model.

precision :: Maybe Double

The overall precision metric value for the trained model.

f1Score :: Maybe Double

The overall F1 score metric for the trained model.

newModelPerformance :: ModelPerformance

Create a value of ModelPerformance with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:recall:ModelPerformance', modelPerformance_recall - The overall recall metric value for the trained model.

$sel:precision:ModelPerformance', modelPerformance_precision - The overall precision metric value for the trained model.

$sel:f1Score:ModelPerformance', modelPerformance_f1Score - The overall F1 score metric for the trained model.

modelPerformance_recall :: Lens' ModelPerformance (Maybe Double)

The overall recall metric value for the trained model.

modelPerformance_precision :: Lens' ModelPerformance (Maybe Double)

The overall precision metric value for the trained model.

modelPerformance_f1Score :: Lens' ModelPerformance (Maybe Double)

The overall F1 score metric for the trained model.

OutputConfig

data OutputConfig

The S3 location where Amazon Lookout for Vision saves model training files.

See: newOutputConfig smart constructor.

Constructors

OutputConfig'

Fields

s3Location :: S3Location

The S3 location for the output.

newOutputConfig

Arguments

:: S3Location
$sel:s3Location:OutputConfig'

-> OutputConfig

Create a value of OutputConfig with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:s3Location:OutputConfig', outputConfig_s3Location - The S3 location for the output.

outputConfig_s3Location :: Lens' OutputConfig S3Location

The S3 location for the output.

OutputS3Object

data OutputS3Object

The S3 location where Amazon Lookout for Vision saves training output.

See: newOutputS3Object smart constructor.

Constructors

OutputS3Object'

Fields

bucket :: Text

The bucket that contains the training output.

key :: Text

The location of the training output in the bucket.

newOutputS3Object

Arguments

:: Text
$sel:bucket:OutputS3Object'

-> Text
$sel:key:OutputS3Object'

-> OutputS3Object

Create a value of OutputS3Object with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:bucket:OutputS3Object', outputS3Object_bucket - The bucket that contains the training output.

$sel:key:OutputS3Object', outputS3Object_key - The location of the training output in the bucket.

outputS3Object_bucket :: Lens' OutputS3Object Text

The bucket that contains the training output.

outputS3Object_key :: Lens' OutputS3Object Text

The location of the training output in the bucket.

ProjectDescription

data ProjectDescription

Describe an Amazon Lookout for Vision project. For more information, see DescribeProject.

See: newProjectDescription smart constructor.

Constructors

ProjectDescription'

Fields

creationTimestamp :: Maybe POSIX

The unix timestamp for the date and time that the project was created.

projectName :: Maybe Text

The name of the project.

projectArn :: Maybe Text

The Amazon Resource Name (ARN) of the project.

datasets :: Maybe [DatasetMetadata]

A list of datasets in the project.

newProjectDescription :: ProjectDescription

Create a value of ProjectDescription with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:creationTimestamp:ProjectDescription', projectDescription_creationTimestamp - The unix timestamp for the date and time that the project was created.

$sel:projectName:ProjectDescription', projectDescription_projectName - The name of the project.

$sel:projectArn:ProjectDescription', projectDescription_projectArn - The Amazon Resource Name (ARN) of the project.

$sel:datasets:ProjectDescription', projectDescription_datasets - A list of datasets in the project.

projectDescription_creationTimestamp :: Lens' ProjectDescription (Maybe UTCTime)

The unix timestamp for the date and time that the project was created.

projectDescription_projectName :: Lens' ProjectDescription (Maybe Text)

The name of the project.

projectDescription_projectArn :: Lens' ProjectDescription (Maybe Text)

The Amazon Resource Name (ARN) of the project.

projectDescription_datasets :: Lens' ProjectDescription (Maybe [DatasetMetadata])

A list of datasets in the project.

ProjectMetadata

data ProjectMetadata

Metadata about an Amazon Lookout for Vision project.

See: newProjectMetadata smart constructor.

Constructors

ProjectMetadata'

Fields

creationTimestamp :: Maybe POSIX

The unix timestamp for the date and time that the project was created.

projectName :: Maybe Text

The name of the project.

projectArn :: Maybe Text

The Amazon Resource Name (ARN) of the project.

newProjectMetadata :: ProjectMetadata

Create a value of ProjectMetadata with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:creationTimestamp:ProjectMetadata', projectMetadata_creationTimestamp - The unix timestamp for the date and time that the project was created.

$sel:projectName:ProjectMetadata', projectMetadata_projectName - The name of the project.

$sel:projectArn:ProjectMetadata', projectMetadata_projectArn - The Amazon Resource Name (ARN) of the project.

projectMetadata_creationTimestamp :: Lens' ProjectMetadata (Maybe UTCTime)

The unix timestamp for the date and time that the project was created.

projectMetadata_projectName :: Lens' ProjectMetadata (Maybe Text)

The name of the project.

projectMetadata_projectArn :: Lens' ProjectMetadata (Maybe Text)

The Amazon Resource Name (ARN) of the project.

S3Location

data S3Location

Information about the location training output.

See: newS3Location smart constructor.

Constructors

S3Location'

Fields

prefix :: Maybe Text

The path of the folder, within the S3 bucket, that contains the training output.

bucket :: Text

The S3 bucket that contains the training output.

newS3Location

Arguments

:: Text
$sel:bucket:S3Location'

-> S3Location

Create a value of S3Location with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:prefix:S3Location', s3Location_prefix - The path of the folder, within the S3 bucket, that contains the training output.

$sel:bucket:S3Location', s3Location_bucket - The S3 bucket that contains the training output.

s3Location_prefix :: Lens' S3Location (Maybe Text)

The path of the folder, within the S3 bucket, that contains the training output.

s3Location_bucket :: Lens' S3Location Text

The S3 bucket that contains the training output.

Tag

data Tag

A key and value pair that is attached to the specified Amazon Lookout for Vision model.

See: newTag smart constructor.

Constructors

Tag'

Fields

key :: Text

The key of the tag that is attached to the specified model.

value :: Text

The value of the tag that is attached to the specified model.

newTag

Arguments

:: Text
$sel:key:Tag'

-> Text
$sel:value:Tag'

-> Tag

Create a value of Tag with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:key:Tag', tag_key - The key of the tag that is attached to the specified model.

$sel:value:Tag', tag_value - The value of the tag that is attached to the specified model.

tag_key :: Lens' Tag Text

The key of the tag that is attached to the specified model.

tag_value :: Lens' Tag Text

The value of the tag that is attached to the specified model.