Amazonka.LookoutVision.Types.DatasetDescription

Description

Documentation

data DatasetDescription

The description for a dataset. For more information, see DescribeDataset.

See: newDatasetDescription smart constructor.

Constructors

DatasetDescription'

Fields

status :: Maybe DatasetStatus

The status of the dataset.

imageStats :: Maybe DatasetImageStats

statusMessage :: Maybe Text

The status message for the dataset.

creationTimestamp :: Maybe POSIX

The Unix timestamp for the time and date that the dataset was created.

datasetType :: Maybe Text

The type of the dataset. The value train represents a training dataset or single dataset project. The value test represents a test dataset.

projectName :: Maybe Text

The name of the project that contains the dataset.

lastUpdatedTimestamp :: Maybe POSIX

The Unix timestamp for the date and time that the dataset was last updated.

newDatasetDescription :: DatasetDescription

Create a value of DatasetDescription with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:status:DatasetDescription', datasetDescription_status - The status of the dataset.

$sel:imageStats:DatasetDescription', datasetDescription_imageStats -

$sel:statusMessage:DatasetDescription', datasetDescription_statusMessage - The status message for the dataset.

$sel:creationTimestamp:DatasetDescription', datasetDescription_creationTimestamp - The Unix timestamp for the time and date that the dataset was created.

$sel:datasetType:DatasetDescription', datasetDescription_datasetType - The type of the dataset. The value train represents a training dataset or single dataset project. The value test represents a test dataset.

$sel:projectName:DatasetDescription', datasetDescription_projectName - The name of the project that contains the dataset.

$sel:lastUpdatedTimestamp:DatasetDescription', datasetDescription_lastUpdatedTimestamp - The Unix timestamp for the date and time that the dataset was last updated.

datasetDescription_status :: Lens' DatasetDescription (Maybe DatasetStatus)

The status of the dataset.

datasetDescription_imageStats :: Lens' DatasetDescription (Maybe DatasetImageStats)

datasetDescription_statusMessage :: Lens' DatasetDescription (Maybe Text)

The status message for the dataset.

datasetDescription_creationTimestamp :: Lens' DatasetDescription (Maybe UTCTime)

The Unix timestamp for the time and date that the dataset was created.

datasetDescription_datasetType :: Lens' DatasetDescription (Maybe Text)

The type of the dataset. The value train represents a training dataset or single dataset project. The value test represents a test dataset.

datasetDescription_projectName :: Lens' DatasetDescription (Maybe Text)

The name of the project that contains the dataset.

datasetDescription_lastUpdatedTimestamp :: Lens' DatasetDescription (Maybe UTCTime)

The Unix timestamp for the date and time that the dataset was last updated.