Amazonka.LookoutVision.Types.ModelDescription

Description

Documentation

data ModelDescription

Describes an Amazon Lookout for Vision model.

See: newModelDescription smart constructor.

Constructors

ModelDescription'

Fields

status :: Maybe ModelStatus

The status of the model.

evaluationResult :: Maybe OutputS3Object

The S3 location where Amazon Lookout for Vision saves the performance metrics.

evaluationEndTimestamp :: Maybe POSIX

The unix timestamp for the date and time that the evaluation ended.

modelArn :: Maybe Text

The Amazon Resource Name (ARN) of the model.

performance :: Maybe ModelPerformance

Performance metrics for the model. Created during training.

kmsKeyId :: Maybe Text

The identifer for the AWS Key Management Service (AWS KMS) key that was used to encrypt the model during training.

statusMessage :: Maybe Text

The status message for the model.

creationTimestamp :: Maybe POSIX

The unix timestamp for the date and time that the model was created.

outputConfig :: Maybe OutputConfig

The S3 location where Amazon Lookout for Vision saves model training files.

modelVersion :: Maybe Text

The version of the model

description :: Maybe Text

The description for the model.

evaluationManifest :: Maybe OutputS3Object

The S3 location where Amazon Lookout for Vision saves the manifest file that was used to test the trained model and generate the performance scores.

newModelDescription :: ModelDescription

Create a value of ModelDescription with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:status:ModelDescription', modelDescription_status - The status of the model.

$sel:evaluationResult:ModelDescription', modelDescription_evaluationResult - The S3 location where Amazon Lookout for Vision saves the performance metrics.

$sel:evaluationEndTimestamp:ModelDescription', modelDescription_evaluationEndTimestamp - The unix timestamp for the date and time that the evaluation ended.

$sel:modelArn:ModelDescription', modelDescription_modelArn - The Amazon Resource Name (ARN) of the model.

$sel:performance:ModelDescription', modelDescription_performance - Performance metrics for the model. Created during training.

$sel:kmsKeyId:ModelDescription', modelDescription_kmsKeyId - The identifer for the AWS Key Management Service (AWS KMS) key that was used to encrypt the model during training.

$sel:statusMessage:ModelDescription', modelDescription_statusMessage - The status message for the model.

$sel:creationTimestamp:ModelDescription', modelDescription_creationTimestamp - The unix timestamp for the date and time that the model was created.

$sel:outputConfig:ModelDescription', modelDescription_outputConfig - The S3 location where Amazon Lookout for Vision saves model training files.

$sel:modelVersion:ModelDescription', modelDescription_modelVersion - The version of the model

$sel:description:ModelDescription', modelDescription_description - The description for the model.

$sel:evaluationManifest:ModelDescription', modelDescription_evaluationManifest - The S3 location where Amazon Lookout for Vision saves the manifest file that was used to test the trained model and generate the performance scores.

modelDescription_status :: Lens' ModelDescription (Maybe ModelStatus)

The status of the model.

modelDescription_evaluationResult :: Lens' ModelDescription (Maybe OutputS3Object)

The S3 location where Amazon Lookout for Vision saves the performance metrics.

modelDescription_evaluationEndTimestamp :: Lens' ModelDescription (Maybe UTCTime)

The unix timestamp for the date and time that the evaluation ended.

modelDescription_modelArn :: Lens' ModelDescription (Maybe Text)

The Amazon Resource Name (ARN) of the model.

modelDescription_performance :: Lens' ModelDescription (Maybe ModelPerformance)

Performance metrics for the model. Created during training.

modelDescription_kmsKeyId :: Lens' ModelDescription (Maybe Text)

The identifer for the AWS Key Management Service (AWS KMS) key that was used to encrypt the model during training.

modelDescription_statusMessage :: Lens' ModelDescription (Maybe Text)

The status message for the model.

modelDescription_creationTimestamp :: Lens' ModelDescription (Maybe UTCTime)

The unix timestamp for the date and time that the model was created.

modelDescription_outputConfig :: Lens' ModelDescription (Maybe OutputConfig)

The S3 location where Amazon Lookout for Vision saves model training files.

modelDescription_modelVersion :: Lens' ModelDescription (Maybe Text)

The version of the model

modelDescription_description :: Lens' ModelDescription (Maybe Text)

The description for the model.

modelDescription_evaluationManifest :: Lens' ModelDescription (Maybe OutputS3Object)

The S3 location where Amazon Lookout for Vision saves the manifest file that was used to test the trained model and generate the performance scores.