Amazonka.LookoutVision.Types.ModelMetadata

Description

Documentation

data ModelMetadata

Describes an Amazon Lookout for Vision model.

See: newModelMetadata smart constructor.

Constructors

ModelMetadata'

Fields

status :: Maybe ModelStatus

The status of the model.

modelArn :: Maybe Text

The Amazon Resource Name (ARN) of the model.

performance :: Maybe ModelPerformance

Performance metrics for the model. Not available until training has successfully completed.

statusMessage :: Maybe Text

The status message for the model.

creationTimestamp :: Maybe POSIX

The unix timestamp for the date and time that the model was created.

modelVersion :: Maybe Text

The version of the model.

description :: Maybe Text

The description for the model.

newModelMetadata :: ModelMetadata

Create a value of ModelMetadata with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:status:ModelMetadata', modelMetadata_status - The status of the model.

$sel:modelArn:ModelMetadata', modelMetadata_modelArn - The Amazon Resource Name (ARN) of the model.

$sel:performance:ModelMetadata', modelMetadata_performance - Performance metrics for the model. Not available until training has successfully completed.

$sel:statusMessage:ModelMetadata', modelMetadata_statusMessage - The status message for the model.

$sel:creationTimestamp:ModelMetadata', modelMetadata_creationTimestamp - The unix timestamp for the date and time that the model was created.

$sel:modelVersion:ModelMetadata', modelMetadata_modelVersion - The version of the model.

$sel:description:ModelMetadata', modelMetadata_description - The description for the model.

modelMetadata_status :: Lens' ModelMetadata (Maybe ModelStatus)

The status of the model.

modelMetadata_modelArn :: Lens' ModelMetadata (Maybe Text)

The Amazon Resource Name (ARN) of the model.

modelMetadata_performance :: Lens' ModelMetadata (Maybe ModelPerformance)

Performance metrics for the model. Not available until training has successfully completed.

modelMetadata_statusMessage :: Lens' ModelMetadata (Maybe Text)

The status message for the model.

modelMetadata_creationTimestamp :: Lens' ModelMetadata (Maybe UTCTime)

The unix timestamp for the date and time that the model was created.

modelMetadata_modelVersion :: Lens' ModelMetadata (Maybe Text)

The version of the model.

modelMetadata_description :: Lens' ModelMetadata (Maybe Text)

The description for the model.