Amazonka.MGN.CreateReplicationConfigurationTemplate

Description

Creates a new ReplicationConfigurationTemplate.

Creating a Request

data CreateReplicationConfigurationTemplate

See: newCreateReplicationConfigurationTemplate smart constructor.

Constructors

CreateReplicationConfigurationTemplate'

Fields

ebsEncryptionKeyArn :: Maybe Text

Request to configure an EBS enryption key during Replication Settings template creation.

tags :: Maybe (Sensitive (HashMap Text Text))

Request to configure tags during Replication Settings template creation.

associateDefaultSecurityGroup :: Bool

Request to associate the default Application Migration Service Security group with the Replication Settings template.

bandwidthThrottling :: Natural

Request to configure bandwidth throttling during Replication Settings template creation.

createPublicIP :: Bool

Request to create Public IP during Replication Settings template creation.

dataPlaneRouting :: ReplicationConfigurationDataPlaneRouting

Request to configure data plane routing during Replication Settings template creation.

defaultLargeStagingDiskType :: ReplicationConfigurationDefaultLargeStagingDiskType

Request to configure the Staging Disk EBS volume type to "gp2" during Replication Settings template creation.

ebsEncryption :: ReplicationConfigurationEbsEncryption

Request to configure EBS enryption during Replication Settings template creation.

replicationServerInstanceType :: Text

Request to configure the Replication Server instance type during Replication Settings template creation.

replicationServersSecurityGroupsIDs :: [Text]

Request to configure the Replication Server Secuirity group ID during Replication Settings template creation.

stagingAreaSubnetId :: Text

Request to configure the Staging Area subnet ID during Replication Settings template creation.

stagingAreaTags :: Sensitive (HashMap Text Text)

Request to configure Staiging Area tags during Replication Settings template creation.

useDedicatedReplicationServer :: Bool

Request to use Dedicated Replication Servers during Replication Settings template creation.

newCreateReplicationConfigurationTemplate

Arguments

:: Bool
$sel:associateDefaultSecurityGroup:CreateReplicationConfigurationTemplate'

-> Natural
$sel:bandwidthThrottling:CreateReplicationConfigurationTemplate'

-> Bool
$sel:createPublicIP:CreateReplicationConfigurationTemplate'

-> ReplicationConfigurationDataPlaneRouting
$sel:dataPlaneRouting:CreateReplicationConfigurationTemplate'

-> ReplicationConfigurationDefaultLargeStagingDiskType
$sel:defaultLargeStagingDiskType:CreateReplicationConfigurationTemplate'

-> ReplicationConfigurationEbsEncryption
$sel:ebsEncryption:CreateReplicationConfigurationTemplate'

-> Text
$sel:replicationServerInstanceType:CreateReplicationConfigurationTemplate'

-> Text
$sel:stagingAreaSubnetId:CreateReplicationConfigurationTemplate'

-> Bool
$sel:useDedicatedReplicationServer:CreateReplicationConfigurationTemplate'

-> CreateReplicationConfigurationTemplate

Create a value of CreateReplicationConfigurationTemplate with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:ebsEncryptionKeyArn:CreateReplicationConfigurationTemplate', createReplicationConfigurationTemplate_ebsEncryptionKeyArn - Request to configure an EBS enryption key during Replication Settings template creation.

$sel:tags:CreateReplicationConfigurationTemplate', createReplicationConfigurationTemplate_tags - Request to configure tags during Replication Settings template creation.

$sel:associateDefaultSecurityGroup:CreateReplicationConfigurationTemplate', createReplicationConfigurationTemplate_associateDefaultSecurityGroup - Request to associate the default Application Migration Service Security group with the Replication Settings template.

$sel:bandwidthThrottling:CreateReplicationConfigurationTemplate', createReplicationConfigurationTemplate_bandwidthThrottling - Request to configure bandwidth throttling during Replication Settings template creation.

$sel:createPublicIP:CreateReplicationConfigurationTemplate', createReplicationConfigurationTemplate_createPublicIP - Request to create Public IP during Replication Settings template creation.

$sel:dataPlaneRouting:CreateReplicationConfigurationTemplate', createReplicationConfigurationTemplate_dataPlaneRouting - Request to configure data plane routing during Replication Settings template creation.

$sel:defaultLargeStagingDiskType:CreateReplicationConfigurationTemplate', createReplicationConfigurationTemplate_defaultLargeStagingDiskType - Request to configure the Staging Disk EBS volume type to "gp2" during Replication Settings template creation.

$sel:ebsEncryption:CreateReplicationConfigurationTemplate', createReplicationConfigurationTemplate_ebsEncryption - Request to configure EBS enryption during Replication Settings template creation.

$sel:replicationServerInstanceType:CreateReplicationConfigurationTemplate', createReplicationConfigurationTemplate_replicationServerInstanceType - Request to configure the Replication Server instance type during Replication Settings template creation.

$sel:replicationServersSecurityGroupsIDs:CreateReplicationConfigurationTemplate', createReplicationConfigurationTemplate_replicationServersSecurityGroupsIDs - Request to configure the Replication Server Secuirity group ID during Replication Settings template creation.

$sel:stagingAreaSubnetId:CreateReplicationConfigurationTemplate', createReplicationConfigurationTemplate_stagingAreaSubnetId - Request to configure the Staging Area subnet ID during Replication Settings template creation.

$sel:stagingAreaTags:CreateReplicationConfigurationTemplate', createReplicationConfigurationTemplate_stagingAreaTags - Request to configure Staiging Area tags during Replication Settings template creation.

$sel:useDedicatedReplicationServer:CreateReplicationConfigurationTemplate', createReplicationConfigurationTemplate_useDedicatedReplicationServer - Request to use Dedicated Replication Servers during Replication Settings template creation.

Request Lenses

createReplicationConfigurationTemplate_ebsEncryptionKeyArn :: Lens' CreateReplicationConfigurationTemplate (Maybe Text)

Request to configure an EBS enryption key during Replication Settings template creation.

createReplicationConfigurationTemplate_tags :: Lens' CreateReplicationConfigurationTemplate (Maybe (HashMap Text Text))

Request to configure tags during Replication Settings template creation.

createReplicationConfigurationTemplate_associateDefaultSecurityGroup :: Lens' CreateReplicationConfigurationTemplate Bool

Request to associate the default Application Migration Service Security group with the Replication Settings template.

createReplicationConfigurationTemplate_bandwidthThrottling :: Lens' CreateReplicationConfigurationTemplate Natural

Request to configure bandwidth throttling during Replication Settings template creation.

createReplicationConfigurationTemplate_createPublicIP :: Lens' CreateReplicationConfigurationTemplate Bool

Request to create Public IP during Replication Settings template creation.

createReplicationConfigurationTemplate_dataPlaneRouting :: Lens' CreateReplicationConfigurationTemplate ReplicationConfigurationDataPlaneRouting

Request to configure data plane routing during Replication Settings template creation.

createReplicationConfigurationTemplate_defaultLargeStagingDiskType :: Lens' CreateReplicationConfigurationTemplate ReplicationConfigurationDefaultLargeStagingDiskType

Request to configure the Staging Disk EBS volume type to "gp2" during Replication Settings template creation.

createReplicationConfigurationTemplate_ebsEncryption :: Lens' CreateReplicationConfigurationTemplate ReplicationConfigurationEbsEncryption

Request to configure EBS enryption during Replication Settings template creation.

createReplicationConfigurationTemplate_replicationServerInstanceType :: Lens' CreateReplicationConfigurationTemplate Text

Request to configure the Replication Server instance type during Replication Settings template creation.

createReplicationConfigurationTemplate_replicationServersSecurityGroupsIDs :: Lens' CreateReplicationConfigurationTemplate [Text]

Request to configure the Replication Server Secuirity group ID during Replication Settings template creation.

createReplicationConfigurationTemplate_stagingAreaSubnetId :: Lens' CreateReplicationConfigurationTemplate Text

Request to configure the Staging Area subnet ID during Replication Settings template creation.

createReplicationConfigurationTemplate_stagingAreaTags :: Lens' CreateReplicationConfigurationTemplate (HashMap Text Text)

Request to configure Staiging Area tags during Replication Settings template creation.

createReplicationConfigurationTemplate_useDedicatedReplicationServer :: Lens' CreateReplicationConfigurationTemplate Bool

Request to use Dedicated Replication Servers during Replication Settings template creation.

Destructuring the Response

data ReplicationConfigurationTemplate

See: newReplicationConfigurationTemplate smart constructor.

Constructors

ReplicationConfigurationTemplate'

Fields

createPublicIP :: Maybe Bool

Replication Configuration template create Public IP.

stagingAreaTags :: Maybe (Sensitive (HashMap Text Text))

Replication Configuration template Staging Area Tags.

arn :: Maybe Text

Replication Configuration template ARN.

stagingAreaSubnetId :: Maybe Text

Replication Configuration template Staging Area subnet ID.

replicationServerInstanceType :: Maybe Text

Replication Configuration template server instance type.

ebsEncryption :: Maybe ReplicationConfigurationEbsEncryption

Replication Configuration template EBS encryption.

associateDefaultSecurityGroup :: Maybe Bool

Replication Configuration template associate default Application Migration Service Security group.

replicationServersSecurityGroupsIDs :: Maybe [Text]

Replication Configuration template server Security Groups IDs.

ebsEncryptionKeyArn :: Maybe Text

Replication Configuration template EBS encryption key ARN.

defaultLargeStagingDiskType :: Maybe ReplicationConfigurationDefaultLargeStagingDiskType

Replication Configuration template use dedault large Staging Disk type.

bandwidthThrottling :: Maybe Natural

Replication Configuration template bandwidth throtting.

dataPlaneRouting :: Maybe ReplicationConfigurationDataPlaneRouting

Replication Configuration template data plane routing.

useDedicatedReplicationServer :: Maybe Bool

Replication Configuration template use Dedicated Replication Server.

tags :: Maybe (Sensitive (HashMap Text Text))

Replication Configuration template Tags.

replicationConfigurationTemplateID :: Text

Replication Configuration template template ID.

newReplicationConfigurationTemplate

Arguments

:: Text
$sel:replicationConfigurationTemplateID:ReplicationConfigurationTemplate'

-> ReplicationConfigurationTemplate

Create a value of ReplicationConfigurationTemplate with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:createPublicIP:ReplicationConfigurationTemplate', replicationConfigurationTemplate_createPublicIP - Replication Configuration template create Public IP.

$sel:stagingAreaTags:ReplicationConfigurationTemplate', replicationConfigurationTemplate_stagingAreaTags - Replication Configuration template Staging Area Tags.

$sel:arn:ReplicationConfigurationTemplate', replicationConfigurationTemplate_arn - Replication Configuration template ARN.

$sel:stagingAreaSubnetId:ReplicationConfigurationTemplate', replicationConfigurationTemplate_stagingAreaSubnetId - Replication Configuration template Staging Area subnet ID.

$sel:replicationServerInstanceType:ReplicationConfigurationTemplate', replicationConfigurationTemplate_replicationServerInstanceType - Replication Configuration template server instance type.

$sel:ebsEncryption:ReplicationConfigurationTemplate', replicationConfigurationTemplate_ebsEncryption - Replication Configuration template EBS encryption.

$sel:associateDefaultSecurityGroup:ReplicationConfigurationTemplate', replicationConfigurationTemplate_associateDefaultSecurityGroup - Replication Configuration template associate default Application Migration Service Security group.

$sel:replicationServersSecurityGroupsIDs:ReplicationConfigurationTemplate', replicationConfigurationTemplate_replicationServersSecurityGroupsIDs - Replication Configuration template server Security Groups IDs.

$sel:ebsEncryptionKeyArn:ReplicationConfigurationTemplate', replicationConfigurationTemplate_ebsEncryptionKeyArn - Replication Configuration template EBS encryption key ARN.

$sel:defaultLargeStagingDiskType:ReplicationConfigurationTemplate', replicationConfigurationTemplate_defaultLargeStagingDiskType - Replication Configuration template use dedault large Staging Disk type.

$sel:bandwidthThrottling:ReplicationConfigurationTemplate', replicationConfigurationTemplate_bandwidthThrottling - Replication Configuration template bandwidth throtting.

$sel:dataPlaneRouting:ReplicationConfigurationTemplate', replicationConfigurationTemplate_dataPlaneRouting - Replication Configuration template data plane routing.

$sel:useDedicatedReplicationServer:ReplicationConfigurationTemplate', replicationConfigurationTemplate_useDedicatedReplicationServer - Replication Configuration template use Dedicated Replication Server.

$sel:tags:ReplicationConfigurationTemplate', replicationConfigurationTemplate_tags - Replication Configuration template Tags.

$sel:replicationConfigurationTemplateID:ReplicationConfigurationTemplate', replicationConfigurationTemplate_replicationConfigurationTemplateID - Replication Configuration template template ID.

Response Lenses

replicationConfigurationTemplate_createPublicIP :: Lens' ReplicationConfigurationTemplate (Maybe Bool)

Replication Configuration template create Public IP.

replicationConfigurationTemplate_stagingAreaTags :: Lens' ReplicationConfigurationTemplate (Maybe (HashMap Text Text))

Replication Configuration template Staging Area Tags.

replicationConfigurationTemplate_arn :: Lens' ReplicationConfigurationTemplate (Maybe Text)

Replication Configuration template ARN.

replicationConfigurationTemplate_stagingAreaSubnetId :: Lens' ReplicationConfigurationTemplate (Maybe Text)

Replication Configuration template Staging Area subnet ID.

replicationConfigurationTemplate_replicationServerInstanceType :: Lens' ReplicationConfigurationTemplate (Maybe Text)

Replication Configuration template server instance type.

replicationConfigurationTemplate_ebsEncryption :: Lens' ReplicationConfigurationTemplate (Maybe ReplicationConfigurationEbsEncryption)

Replication Configuration template EBS encryption.

replicationConfigurationTemplate_associateDefaultSecurityGroup :: Lens' ReplicationConfigurationTemplate (Maybe Bool)

Replication Configuration template associate default Application Migration Service Security group.

replicationConfigurationTemplate_replicationServersSecurityGroupsIDs :: Lens' ReplicationConfigurationTemplate (Maybe [Text])

Replication Configuration template server Security Groups IDs.

replicationConfigurationTemplate_ebsEncryptionKeyArn :: Lens' ReplicationConfigurationTemplate (Maybe Text)

Replication Configuration template EBS encryption key ARN.

replicationConfigurationTemplate_defaultLargeStagingDiskType :: Lens' ReplicationConfigurationTemplate (Maybe ReplicationConfigurationDefaultLargeStagingDiskType)

Replication Configuration template use dedault large Staging Disk type.

replicationConfigurationTemplate_bandwidthThrottling :: Lens' ReplicationConfigurationTemplate (Maybe Natural)

Replication Configuration template bandwidth throtting.

replicationConfigurationTemplate_dataPlaneRouting :: Lens' ReplicationConfigurationTemplate (Maybe ReplicationConfigurationDataPlaneRouting)

Replication Configuration template data plane routing.

replicationConfigurationTemplate_useDedicatedReplicationServer :: Lens' ReplicationConfigurationTemplate (Maybe Bool)

Replication Configuration template use Dedicated Replication Server.

replicationConfigurationTemplate_tags :: Lens' ReplicationConfigurationTemplate (Maybe (HashMap Text Text))

Replication Configuration template Tags.

replicationConfigurationTemplate_replicationConfigurationTemplateID :: Lens' ReplicationConfigurationTemplate Text

Replication Configuration template template ID.