Amazonka.MGN.Lens

Description

Operations

UpdateLaunchConfiguration

updateLaunchConfiguration_targetInstanceTypeRightSizingMethod :: Lens' UpdateLaunchConfiguration (Maybe TargetInstanceTypeRightSizingMethod)

Update Launch configuration Target instance right sizing request.

updateLaunchConfiguration_launchDisposition :: Lens' UpdateLaunchConfiguration (Maybe LaunchDisposition)

Update Launch configuration launch disposition request.

updateLaunchConfiguration_copyTags :: Lens' UpdateLaunchConfiguration (Maybe Bool)

Update Launch configuration copy Tags request.

updateLaunchConfiguration_name :: Lens' UpdateLaunchConfiguration (Maybe Text)

Update Launch configuration name request.

updateLaunchConfiguration_licensing :: Lens' UpdateLaunchConfiguration (Maybe Licensing)

Update Launch configuration licensing request.

updateLaunchConfiguration_copyPrivateIp :: Lens' UpdateLaunchConfiguration (Maybe Bool)

Update Launch configuration copy Private IP request.

updateLaunchConfiguration_sourceServerID :: Lens' UpdateLaunchConfiguration Text

Update Launch configuration by Source Server ID request.

launchConfiguration_ec2LaunchTemplateID :: Lens' LaunchConfiguration (Maybe Text)

Configure EC2 lauch configuration template ID.

launchConfiguration_targetInstanceTypeRightSizingMethod :: Lens' LaunchConfiguration (Maybe TargetInstanceTypeRightSizingMethod)

Configure launch configuration Target instance type right sizing method.

launchConfiguration_launchDisposition :: Lens' LaunchConfiguration (Maybe LaunchDisposition)

Configure launch dispostion for launch configuration.

launchConfiguration_copyTags :: Lens' LaunchConfiguration (Maybe Bool)

Copy Tags during Launch Configuration.

launchConfiguration_name :: Lens' LaunchConfiguration (Maybe Text)

Configure launch configuration name.

launchConfiguration_sourceServerID :: Lens' LaunchConfiguration (Maybe Text)

Configure launch configuration Source Server ID.

launchConfiguration_licensing :: Lens' LaunchConfiguration (Maybe Licensing)

Configure launch configuration OS licensing.

launchConfiguration_copyPrivateIp :: Lens' LaunchConfiguration (Maybe Bool)

Copy Private IP during Launch Configuration.

DescribeReplicationConfigurationTemplates

describeReplicationConfigurationTemplates_nextToken :: Lens' DescribeReplicationConfigurationTemplates (Maybe Text)

Request to describe Replication Configuration template by next token.

describeReplicationConfigurationTemplates_maxResults :: Lens' DescribeReplicationConfigurationTemplates (Maybe Natural)

Request to describe Replication Configuration template by max results.

describeReplicationConfigurationTemplates_replicationConfigurationTemplateIDs :: Lens' DescribeReplicationConfigurationTemplates [Text]

Request to describe Replication Configuration template by template IDs.

describeReplicationConfigurationTemplatesResponse_items :: Lens' DescribeReplicationConfigurationTemplatesResponse (Maybe [ReplicationConfigurationTemplate])

Request to describe Replication Configuration template by items.

describeReplicationConfigurationTemplatesResponse_nextToken :: Lens' DescribeReplicationConfigurationTemplatesResponse (Maybe Text)

Request to describe Replication Configuration template by next token.

describeReplicationConfigurationTemplatesResponse_httpStatus :: Lens' DescribeReplicationConfigurationTemplatesResponse Int

The response's http status code.

ListTagsForResource

listTagsForResource_resourceArn :: Lens' ListTagsForResource Text

List tags for resource request by ARN.

listTagsForResourceResponse_tags :: Lens' ListTagsForResourceResponse (Maybe (HashMap Text Text))

List tags for resource response.

listTagsForResourceResponse_httpStatus :: Lens' ListTagsForResourceResponse Int

The response's http status code.

InitializeService

initializeServiceResponse_httpStatus :: Lens' InitializeServiceResponse Int

The response's http status code.

UpdateReplicationConfigurationTemplate

updateReplicationConfigurationTemplate_createPublicIP :: Lens' UpdateReplicationConfigurationTemplate (Maybe Bool)

Update replication configuration template create Public IP request.

updateReplicationConfigurationTemplate_stagingAreaTags :: Lens' UpdateReplicationConfigurationTemplate (Maybe (HashMap Text Text))

Update replication configuration template Staging Area Tags request.

updateReplicationConfigurationTemplate_arn :: Lens' UpdateReplicationConfigurationTemplate (Maybe Text)

Update replication configuration template ARN request.

updateReplicationConfigurationTemplate_stagingAreaSubnetId :: Lens' UpdateReplicationConfigurationTemplate (Maybe Text)

Update replication configuration template Staging Area subnet ID request.

updateReplicationConfigurationTemplate_replicationServerInstanceType :: Lens' UpdateReplicationConfigurationTemplate (Maybe Text)

Update replication configuration template Replication Server instance type request.

updateReplicationConfigurationTemplate_ebsEncryption :: Lens' UpdateReplicationConfigurationTemplate (Maybe ReplicationConfigurationEbsEncryption)

Update replication configuration template EBS encryption request.

updateReplicationConfigurationTemplate_associateDefaultSecurityGroup :: Lens' UpdateReplicationConfigurationTemplate (Maybe Bool)

Update replication configuration template associate default Application Migration Service Security group request.

updateReplicationConfigurationTemplate_replicationServersSecurityGroupsIDs :: Lens' UpdateReplicationConfigurationTemplate (Maybe [Text])

Update replication configuration template Replication Server Security groups IDs request.

updateReplicationConfigurationTemplate_ebsEncryptionKeyArn :: Lens' UpdateReplicationConfigurationTemplate (Maybe Text)

Update replication configuration template EBS encryption key ARN request.

updateReplicationConfigurationTemplate_defaultLargeStagingDiskType :: Lens' UpdateReplicationConfigurationTemplate (Maybe ReplicationConfigurationDefaultLargeStagingDiskType)

Update replication configuration template use default large Staging Disk type request.

updateReplicationConfigurationTemplate_bandwidthThrottling :: Lens' UpdateReplicationConfigurationTemplate (Maybe Natural)

Update replication configuration template bandwidth throttling request.

updateReplicationConfigurationTemplate_dataPlaneRouting :: Lens' UpdateReplicationConfigurationTemplate (Maybe ReplicationConfigurationDataPlaneRouting)

Update replication configuration template data plane routing request.

updateReplicationConfigurationTemplate_useDedicatedReplicationServer :: Lens' UpdateReplicationConfigurationTemplate (Maybe Bool)

Update replication configuration template use dedicated Replication Server request.

updateReplicationConfigurationTemplate_replicationConfigurationTemplateID :: Lens' UpdateReplicationConfigurationTemplate Text

Update replication configuration template template ID request.

replicationConfigurationTemplate_createPublicIP :: Lens' ReplicationConfigurationTemplate (Maybe Bool)

Replication Configuration template create Public IP.

replicationConfigurationTemplate_stagingAreaTags :: Lens' ReplicationConfigurationTemplate (Maybe (HashMap Text Text))

Replication Configuration template Staging Area Tags.

replicationConfigurationTemplate_arn :: Lens' ReplicationConfigurationTemplate (Maybe Text)

Replication Configuration template ARN.

replicationConfigurationTemplate_stagingAreaSubnetId :: Lens' ReplicationConfigurationTemplate (Maybe Text)

Replication Configuration template Staging Area subnet ID.

replicationConfigurationTemplate_replicationServerInstanceType :: Lens' ReplicationConfigurationTemplate (Maybe Text)

Replication Configuration template server instance type.

replicationConfigurationTemplate_ebsEncryption :: Lens' ReplicationConfigurationTemplate (Maybe ReplicationConfigurationEbsEncryption)

Replication Configuration template EBS encryption.

replicationConfigurationTemplate_associateDefaultSecurityGroup :: Lens' ReplicationConfigurationTemplate (Maybe Bool)

Replication Configuration template associate default Application Migration Service Security group.

replicationConfigurationTemplate_replicationServersSecurityGroupsIDs :: Lens' ReplicationConfigurationTemplate (Maybe [Text])

Replication Configuration template server Security Groups IDs.

replicationConfigurationTemplate_ebsEncryptionKeyArn :: Lens' ReplicationConfigurationTemplate (Maybe Text)

Replication Configuration template EBS encryption key ARN.

replicationConfigurationTemplate_defaultLargeStagingDiskType :: Lens' ReplicationConfigurationTemplate (Maybe ReplicationConfigurationDefaultLargeStagingDiskType)

Replication Configuration template use dedault large Staging Disk type.

replicationConfigurationTemplate_bandwidthThrottling :: Lens' ReplicationConfigurationTemplate (Maybe Natural)

Replication Configuration template bandwidth throtting.

replicationConfigurationTemplate_dataPlaneRouting :: Lens' ReplicationConfigurationTemplate (Maybe ReplicationConfigurationDataPlaneRouting)

Replication Configuration template data plane routing.

replicationConfigurationTemplate_useDedicatedReplicationServer :: Lens' ReplicationConfigurationTemplate (Maybe Bool)

Replication Configuration template use Dedicated Replication Server.

replicationConfigurationTemplate_tags :: Lens' ReplicationConfigurationTemplate (Maybe (HashMap Text Text))

Replication Configuration template Tags.

replicationConfigurationTemplate_replicationConfigurationTemplateID :: Lens' ReplicationConfigurationTemplate Text

Replication Configuration template template ID.

DeleteReplicationConfigurationTemplate

deleteReplicationConfigurationTemplate_replicationConfigurationTemplateID :: Lens' DeleteReplicationConfigurationTemplate Text

Request to delete Replication Configuration Template from service by Replication Configuration Template ID.

deleteReplicationConfigurationTemplateResponse_httpStatus :: Lens' DeleteReplicationConfigurationTemplateResponse Int

The response's http status code.

CreateReplicationConfigurationTemplate

createReplicationConfigurationTemplate_ebsEncryptionKeyArn :: Lens' CreateReplicationConfigurationTemplate (Maybe Text)

Request to configure an EBS enryption key during Replication Settings template creation.

createReplicationConfigurationTemplate_tags :: Lens' CreateReplicationConfigurationTemplate (Maybe (HashMap Text Text))

Request to configure tags during Replication Settings template creation.

createReplicationConfigurationTemplate_associateDefaultSecurityGroup :: Lens' CreateReplicationConfigurationTemplate Bool

Request to associate the default Application Migration Service Security group with the Replication Settings template.

createReplicationConfigurationTemplate_bandwidthThrottling :: Lens' CreateReplicationConfigurationTemplate Natural

Request to configure bandwidth throttling during Replication Settings template creation.

createReplicationConfigurationTemplate_createPublicIP :: Lens' CreateReplicationConfigurationTemplate Bool

Request to create Public IP during Replication Settings template creation.

createReplicationConfigurationTemplate_dataPlaneRouting :: Lens' CreateReplicationConfigurationTemplate ReplicationConfigurationDataPlaneRouting

Request to configure data plane routing during Replication Settings template creation.

createReplicationConfigurationTemplate_defaultLargeStagingDiskType :: Lens' CreateReplicationConfigurationTemplate ReplicationConfigurationDefaultLargeStagingDiskType

Request to configure the Staging Disk EBS volume type to "gp2" during Replication Settings template creation.

createReplicationConfigurationTemplate_ebsEncryption :: Lens' CreateReplicationConfigurationTemplate ReplicationConfigurationEbsEncryption

Request to configure EBS enryption during Replication Settings template creation.

createReplicationConfigurationTemplate_replicationServerInstanceType :: Lens' CreateReplicationConfigurationTemplate Text

Request to configure the Replication Server instance type during Replication Settings template creation.

createReplicationConfigurationTemplate_replicationServersSecurityGroupsIDs :: Lens' CreateReplicationConfigurationTemplate [Text]

Request to configure the Replication Server Secuirity group ID during Replication Settings template creation.

createReplicationConfigurationTemplate_stagingAreaSubnetId :: Lens' CreateReplicationConfigurationTemplate Text

Request to configure the Staging Area subnet ID during Replication Settings template creation.

createReplicationConfigurationTemplate_stagingAreaTags :: Lens' CreateReplicationConfigurationTemplate (HashMap Text Text)

Request to configure Staiging Area tags during Replication Settings template creation.

createReplicationConfigurationTemplate_useDedicatedReplicationServer :: Lens' CreateReplicationConfigurationTemplate Bool

Request to use Dedicated Replication Servers during Replication Settings template creation.

replicationConfigurationTemplate_createPublicIP :: Lens' ReplicationConfigurationTemplate (Maybe Bool)

Replication Configuration template create Public IP.

replicationConfigurationTemplate_stagingAreaTags :: Lens' ReplicationConfigurationTemplate (Maybe (HashMap Text Text))

Replication Configuration template Staging Area Tags.

replicationConfigurationTemplate_arn :: Lens' ReplicationConfigurationTemplate (Maybe Text)

Replication Configuration template ARN.

replicationConfigurationTemplate_stagingAreaSubnetId :: Lens' ReplicationConfigurationTemplate (Maybe Text)

Replication Configuration template Staging Area subnet ID.

replicationConfigurationTemplate_replicationServerInstanceType :: Lens' ReplicationConfigurationTemplate (Maybe Text)

Replication Configuration template server instance type.

replicationConfigurationTemplate_ebsEncryption :: Lens' ReplicationConfigurationTemplate (Maybe ReplicationConfigurationEbsEncryption)

Replication Configuration template EBS encryption.

replicationConfigurationTemplate_associateDefaultSecurityGroup :: Lens' ReplicationConfigurationTemplate (Maybe Bool)

Replication Configuration template associate default Application Migration Service Security group.

replicationConfigurationTemplate_replicationServersSecurityGroupsIDs :: Lens' ReplicationConfigurationTemplate (Maybe [Text])

Replication Configuration template server Security Groups IDs.

replicationConfigurationTemplate_ebsEncryptionKeyArn :: Lens' ReplicationConfigurationTemplate (Maybe Text)

Replication Configuration template EBS encryption key ARN.

replicationConfigurationTemplate_defaultLargeStagingDiskType :: Lens' ReplicationConfigurationTemplate (Maybe ReplicationConfigurationDefaultLargeStagingDiskType)

Replication Configuration template use dedault large Staging Disk type.

replicationConfigurationTemplate_bandwidthThrottling :: Lens' ReplicationConfigurationTemplate (Maybe Natural)

Replication Configuration template bandwidth throtting.

replicationConfigurationTemplate_dataPlaneRouting :: Lens' ReplicationConfigurationTemplate (Maybe ReplicationConfigurationDataPlaneRouting)

Replication Configuration template data plane routing.

replicationConfigurationTemplate_useDedicatedReplicationServer :: Lens' ReplicationConfigurationTemplate (Maybe Bool)

Replication Configuration template use Dedicated Replication Server.

replicationConfigurationTemplate_tags :: Lens' ReplicationConfigurationTemplate (Maybe (HashMap Text Text))

Replication Configuration template Tags.

replicationConfigurationTemplate_replicationConfigurationTemplateID :: Lens' ReplicationConfigurationTemplate Text

Replication Configuration template template ID.

DescribeJobLogItems

describeJobLogItems_nextToken :: Lens' DescribeJobLogItems (Maybe Text)

Request to describe Job log next token.

describeJobLogItems_maxResults :: Lens' DescribeJobLogItems (Maybe Natural)

Request to describe Job log item maximum results.

describeJobLogItems_jobID :: Lens' DescribeJobLogItems Text

Request to describe Job log job ID.

describeJobLogItemsResponse_items :: Lens' DescribeJobLogItemsResponse (Maybe [JobLog])

Request to describe Job log response items.

describeJobLogItemsResponse_nextToken :: Lens' DescribeJobLogItemsResponse (Maybe Text)

Request to describe Job log response next token.

describeJobLogItemsResponse_httpStatus :: Lens' DescribeJobLogItemsResponse Int

The response's http status code.

DisconnectFromService

disconnectFromService_sourceServerID :: Lens' DisconnectFromService Text

Request to disconnect Source Server from service by Server ID.

sourceServer_sourceProperties :: Lens' SourceServer (Maybe SourceProperties)

Source server properties.

sourceServer_arn :: Lens' SourceServer (Maybe Text)

Source server ARN.

sourceServer_launchedInstance :: Lens' SourceServer (Maybe LaunchedInstance)

Source server launched instance.

sourceServer_lifeCycle :: Lens' SourceServer (Maybe LifeCycle)

Source server lifecycle state.

sourceServer_isArchived :: Lens' SourceServer (Maybe Bool)

Source server archived status.

sourceServer_dataReplicationInfo :: Lens' SourceServer (Maybe DataReplicationInfo)

Source server data replication info.

sourceServer_sourceServerID :: Lens' SourceServer (Maybe Text)

Source server ID.

sourceServer_tags :: Lens' SourceServer (Maybe (HashMap Text Text))

Source server Tags.

StartTest

startTest_tags :: Lens' StartTest (Maybe (HashMap Text Text))

Start Test by Tags.

startTest_sourceServerIDs :: Lens' StartTest (NonEmpty Text)

Start Test for Source Server IDs.

startTestResponse_job :: Lens' StartTestResponse (Maybe Job)

Start Test Job response.

startTestResponse_httpStatus :: Lens' StartTestResponse Int

The response's http status code.

DescribeSourceServers

describeSourceServers_nextToken :: Lens' DescribeSourceServers (Maybe Text)

Request to filter Source Servers list by next token.

describeSourceServers_maxResults :: Lens' DescribeSourceServers (Maybe Natural)

Request to filter Source Servers list by maximum results.

describeSourceServers_filters :: Lens' DescribeSourceServers DescribeSourceServersRequestFilters

Request to filter Source Servers list.

describeSourceServersResponse_items :: Lens' DescribeSourceServersResponse (Maybe [SourceServer])

Request to filter Source Servers list by item.

describeSourceServersResponse_nextToken :: Lens' DescribeSourceServersResponse (Maybe Text)

Request to filter Source Servers next token.

describeSourceServersResponse_httpStatus :: Lens' DescribeSourceServersResponse Int

The response's http status code.

DeleteJob

deleteJob_jobID :: Lens' DeleteJob Text

Request to delete Job from service by Job ID.

deleteJobResponse_httpStatus :: Lens' DeleteJobResponse Int

The response's http status code.

FinalizeCutover

finalizeCutover_sourceServerID :: Lens' FinalizeCutover Text

Request to finalize Cutover by Soure Server ID.

sourceServer_sourceProperties :: Lens' SourceServer (Maybe SourceProperties)

Source server properties.

sourceServer_arn :: Lens' SourceServer (Maybe Text)

Source server ARN.

sourceServer_launchedInstance :: Lens' SourceServer (Maybe LaunchedInstance)

Source server launched instance.

sourceServer_lifeCycle :: Lens' SourceServer (Maybe LifeCycle)

Source server lifecycle state.

sourceServer_isArchived :: Lens' SourceServer (Maybe Bool)

Source server archived status.

sourceServer_dataReplicationInfo :: Lens' SourceServer (Maybe DataReplicationInfo)

Source server data replication info.

sourceServer_sourceServerID :: Lens' SourceServer (Maybe Text)

Source server ID.

sourceServer_tags :: Lens' SourceServer (Maybe (HashMap Text Text))

Source server Tags.

DescribeJobs

describeJobs_nextToken :: Lens' DescribeJobs (Maybe Text)

Request to describe Job logby next token.

describeJobs_maxResults :: Lens' DescribeJobs (Maybe Natural)

Request to describe Job log by max results.

describeJobs_filters :: Lens' DescribeJobs DescribeJobsRequestFilters

Request to describe Job log filters.

describeJobsResponse_items :: Lens' DescribeJobsResponse (Maybe [Job])

Request to describe Job log items.

describeJobsResponse_nextToken :: Lens' DescribeJobsResponse (Maybe Text)

Request to describe Job response by next token.

describeJobsResponse_httpStatus :: Lens' DescribeJobsResponse Int

The response's http status code.

MarkAsArchived

markAsArchived_sourceServerID :: Lens' MarkAsArchived Text

Mark as archived by Source Server ID.

sourceServer_sourceProperties :: Lens' SourceServer (Maybe SourceProperties)

Source server properties.

sourceServer_arn :: Lens' SourceServer (Maybe Text)

Source server ARN.

sourceServer_launchedInstance :: Lens' SourceServer (Maybe LaunchedInstance)

Source server launched instance.

sourceServer_lifeCycle :: Lens' SourceServer (Maybe LifeCycle)

Source server lifecycle state.

sourceServer_isArchived :: Lens' SourceServer (Maybe Bool)

Source server archived status.

sourceServer_dataReplicationInfo :: Lens' SourceServer (Maybe DataReplicationInfo)

Source server data replication info.

sourceServer_sourceServerID :: Lens' SourceServer (Maybe Text)

Source server ID.

sourceServer_tags :: Lens' SourceServer (Maybe (HashMap Text Text))

Source server Tags.

StartCutover

startCutover_tags :: Lens' StartCutover (Maybe (HashMap Text Text))

Start Cutover by Tags.

startCutover_sourceServerIDs :: Lens' StartCutover (NonEmpty Text)

Start Cutover by Source Server IDs.

startCutoverResponse_job :: Lens' StartCutoverResponse (Maybe Job)

Start Cutover Job response.

startCutoverResponse_httpStatus :: Lens' StartCutoverResponse Int

The response's http status code.

RetryDataReplication

retryDataReplication_sourceServerID :: Lens' RetryDataReplication Text

Retry data replication for Source Server ID.

sourceServer_sourceProperties :: Lens' SourceServer (Maybe SourceProperties)

Source server properties.

sourceServer_arn :: Lens' SourceServer (Maybe Text)

Source server ARN.

sourceServer_launchedInstance :: Lens' SourceServer (Maybe LaunchedInstance)

Source server launched instance.

sourceServer_lifeCycle :: Lens' SourceServer (Maybe LifeCycle)

Source server lifecycle state.

sourceServer_isArchived :: Lens' SourceServer (Maybe Bool)

Source server archived status.

sourceServer_dataReplicationInfo :: Lens' SourceServer (Maybe DataReplicationInfo)

Source server data replication info.

sourceServer_sourceServerID :: Lens' SourceServer (Maybe Text)

Source server ID.

sourceServer_tags :: Lens' SourceServer (Maybe (HashMap Text Text))

Source server Tags.

GetReplicationConfiguration

getReplicationConfiguration_sourceServerID :: Lens' GetReplicationConfiguration Text

Request to get Replication Configuaration by Source Server ID.

replicationConfiguration_createPublicIP :: Lens' ReplicationConfiguration (Maybe Bool)

Replication Configuration create Public IP.

replicationConfiguration_stagingAreaTags :: Lens' ReplicationConfiguration (Maybe (HashMap Text Text))

Replication Configuration Staging Area tags.

replicationConfiguration_stagingAreaSubnetId :: Lens' ReplicationConfiguration (Maybe Text)

Replication Configuration Staging Area subnet ID.

replicationConfiguration_replicationServerInstanceType :: Lens' ReplicationConfiguration (Maybe Text)

Replication Configuration Replication Server instance type.

replicationConfiguration_ebsEncryption :: Lens' ReplicationConfiguration (Maybe ReplicationConfigurationEbsEncryption)

Replication Configuration EBS encryption.

replicationConfiguration_associateDefaultSecurityGroup :: Lens' ReplicationConfiguration (Maybe Bool)

Replication Configuration associate default Application Migration Service Security Group.

replicationConfiguration_replicationServersSecurityGroupsIDs :: Lens' ReplicationConfiguration (Maybe [Text])

Replication Configuration Replication Server Security Group IDs.

replicationConfiguration_ebsEncryptionKeyArn :: Lens' ReplicationConfiguration (Maybe Text)

Replication Configuration EBS encryption key ARN.

replicationConfiguration_name :: Lens' ReplicationConfiguration (Maybe Text)

Replication Configuration name.

replicationConfiguration_defaultLargeStagingDiskType :: Lens' ReplicationConfiguration (Maybe ReplicationConfigurationDefaultLargeStagingDiskType)

Replication Configuration use default large Staging Disks.

replicationConfiguration_bandwidthThrottling :: Lens' ReplicationConfiguration (Maybe Natural)

Replication Configuration set bandwidth throttling.

replicationConfiguration_dataPlaneRouting :: Lens' ReplicationConfiguration (Maybe ReplicationConfigurationDataPlaneRouting)

Replication Configuration data plane routing.

replicationConfiguration_sourceServerID :: Lens' ReplicationConfiguration (Maybe Text)

Replication Configuration Source Server ID.

replicationConfiguration_replicatedDisks :: Lens' ReplicationConfiguration (Maybe [ReplicationConfigurationReplicatedDisk])

Replication Configuration replicated disks.

replicationConfiguration_useDedicatedReplicationServer :: Lens' ReplicationConfiguration (Maybe Bool)

Replication Configuration use Dedicated Replication Server.

ChangeServerLifeCycleState

changeServerLifeCycleState_lifeCycle :: Lens' ChangeServerLifeCycleState ChangeServerLifeCycleStateSourceServerLifecycle

The request to change the source server migration lifecycle state.

changeServerLifeCycleState_sourceServerID :: Lens' ChangeServerLifeCycleState Text

The request to change the source server migration lifecycle state by source server ID.

sourceServer_sourceProperties :: Lens' SourceServer (Maybe SourceProperties)

Source server properties.

sourceServer_arn :: Lens' SourceServer (Maybe Text)

Source server ARN.

sourceServer_launchedInstance :: Lens' SourceServer (Maybe LaunchedInstance)

Source server launched instance.

sourceServer_lifeCycle :: Lens' SourceServer (Maybe LifeCycle)

Source server lifecycle state.

sourceServer_isArchived :: Lens' SourceServer (Maybe Bool)

Source server archived status.

sourceServer_dataReplicationInfo :: Lens' SourceServer (Maybe DataReplicationInfo)

Source server data replication info.

sourceServer_sourceServerID :: Lens' SourceServer (Maybe Text)

Source server ID.

sourceServer_tags :: Lens' SourceServer (Maybe (HashMap Text Text))

Source server Tags.

TerminateTargetInstances

terminateTargetInstances_tags :: Lens' TerminateTargetInstances (Maybe (HashMap Text Text))

Terminate Target instance by Tags.

terminateTargetInstances_sourceServerIDs :: Lens' TerminateTargetInstances (NonEmpty Text)

Terminate Target instance by Source Server IDs.

terminateTargetInstancesResponse_job :: Lens' TerminateTargetInstancesResponse (Maybe Job)

Terminate Target instance Job response.

terminateTargetInstancesResponse_httpStatus :: Lens' TerminateTargetInstancesResponse Int

The response's http status code.

UpdateReplicationConfiguration

updateReplicationConfiguration_createPublicIP :: Lens' UpdateReplicationConfiguration (Maybe Bool)

Update replication configuration create Public IP request.

updateReplicationConfiguration_stagingAreaTags :: Lens' UpdateReplicationConfiguration (Maybe (HashMap Text Text))

Update replication configuration Staging Area Tags request.

updateReplicationConfiguration_stagingAreaSubnetId :: Lens' UpdateReplicationConfiguration (Maybe Text)

Update replication configuration Staging Area subnet request.

updateReplicationConfiguration_replicationServerInstanceType :: Lens' UpdateReplicationConfiguration (Maybe Text)

Update replication configuration Replication Server instance type request.

updateReplicationConfiguration_ebsEncryption :: Lens' UpdateReplicationConfiguration (Maybe ReplicationConfigurationEbsEncryption)

Update replication configuration EBS encryption request.

updateReplicationConfiguration_associateDefaultSecurityGroup :: Lens' UpdateReplicationConfiguration (Maybe Bool)

Update replication configuration associate default Application Migration Service Security group request.

updateReplicationConfiguration_replicationServersSecurityGroupsIDs :: Lens' UpdateReplicationConfiguration (Maybe [Text])

Update replication configuration Replication Server Security Groups IDs request.

updateReplicationConfiguration_ebsEncryptionKeyArn :: Lens' UpdateReplicationConfiguration (Maybe Text)

Update replication configuration EBS encryption key ARN request.

updateReplicationConfiguration_name :: Lens' UpdateReplicationConfiguration (Maybe Text)

Update replication configuration name request.

updateReplicationConfiguration_defaultLargeStagingDiskType :: Lens' UpdateReplicationConfiguration (Maybe ReplicationConfigurationDefaultLargeStagingDiskType)

Update replication configuration use default large Staging Disk type request.

updateReplicationConfiguration_bandwidthThrottling :: Lens' UpdateReplicationConfiguration (Maybe Natural)

Update replication configuration bandwidth throttling request.

updateReplicationConfiguration_dataPlaneRouting :: Lens' UpdateReplicationConfiguration (Maybe ReplicationConfigurationDataPlaneRouting)

Update replication configuration data plane routing request.

updateReplicationConfiguration_replicatedDisks :: Lens' UpdateReplicationConfiguration (Maybe [ReplicationConfigurationReplicatedDisk])

Update replication configuration replicated disks request.

updateReplicationConfiguration_useDedicatedReplicationServer :: Lens' UpdateReplicationConfiguration (Maybe Bool)

Update replication configuration use dedicated Replication Server request.

updateReplicationConfiguration_sourceServerID :: Lens' UpdateReplicationConfiguration Text

Update replication configuration Source Server ID request.

replicationConfiguration_createPublicIP :: Lens' ReplicationConfiguration (Maybe Bool)

Replication Configuration create Public IP.

replicationConfiguration_stagingAreaTags :: Lens' ReplicationConfiguration (Maybe (HashMap Text Text))

Replication Configuration Staging Area tags.

replicationConfiguration_stagingAreaSubnetId :: Lens' ReplicationConfiguration (Maybe Text)

Replication Configuration Staging Area subnet ID.

replicationConfiguration_replicationServerInstanceType :: Lens' ReplicationConfiguration (Maybe Text)

Replication Configuration Replication Server instance type.

replicationConfiguration_ebsEncryption :: Lens' ReplicationConfiguration (Maybe ReplicationConfigurationEbsEncryption)

Replication Configuration EBS encryption.

replicationConfiguration_associateDefaultSecurityGroup :: Lens' ReplicationConfiguration (Maybe Bool)

Replication Configuration associate default Application Migration Service Security Group.

replicationConfiguration_replicationServersSecurityGroupsIDs :: Lens' ReplicationConfiguration (Maybe [Text])

Replication Configuration Replication Server Security Group IDs.

replicationConfiguration_ebsEncryptionKeyArn :: Lens' ReplicationConfiguration (Maybe Text)

Replication Configuration EBS encryption key ARN.

replicationConfiguration_name :: Lens' ReplicationConfiguration (Maybe Text)

Replication Configuration name.

replicationConfiguration_defaultLargeStagingDiskType :: Lens' ReplicationConfiguration (Maybe ReplicationConfigurationDefaultLargeStagingDiskType)

Replication Configuration use default large Staging Disks.

replicationConfiguration_bandwidthThrottling :: Lens' ReplicationConfiguration (Maybe Natural)

Replication Configuration set bandwidth throttling.

replicationConfiguration_dataPlaneRouting :: Lens' ReplicationConfiguration (Maybe ReplicationConfigurationDataPlaneRouting)

Replication Configuration data plane routing.

replicationConfiguration_sourceServerID :: Lens' ReplicationConfiguration (Maybe Text)

Replication Configuration Source Server ID.

replicationConfiguration_replicatedDisks :: Lens' ReplicationConfiguration (Maybe [ReplicationConfigurationReplicatedDisk])

Replication Configuration replicated disks.

replicationConfiguration_useDedicatedReplicationServer :: Lens' ReplicationConfiguration (Maybe Bool)

Replication Configuration use Dedicated Replication Server.

TagResource

tagResource_resourceArn :: Lens' TagResource Text

Tag resource by ARN.

tagResource_tags :: Lens' TagResource (HashMap Text Text)

Tag resource by Tags.

GetLaunchConfiguration

getLaunchConfiguration_sourceServerID :: Lens' GetLaunchConfiguration Text

Request to get Launch Configuration information by Source Server ID.

launchConfiguration_ec2LaunchTemplateID :: Lens' LaunchConfiguration (Maybe Text)

Configure EC2 lauch configuration template ID.

launchConfiguration_targetInstanceTypeRightSizingMethod :: Lens' LaunchConfiguration (Maybe TargetInstanceTypeRightSizingMethod)

Configure launch configuration Target instance type right sizing method.

launchConfiguration_launchDisposition :: Lens' LaunchConfiguration (Maybe LaunchDisposition)

Configure launch dispostion for launch configuration.

launchConfiguration_copyTags :: Lens' LaunchConfiguration (Maybe Bool)

Copy Tags during Launch Configuration.

launchConfiguration_name :: Lens' LaunchConfiguration (Maybe Text)

Configure launch configuration name.

launchConfiguration_sourceServerID :: Lens' LaunchConfiguration (Maybe Text)

Configure launch configuration Source Server ID.

launchConfiguration_licensing :: Lens' LaunchConfiguration (Maybe Licensing)

Configure launch configuration OS licensing.

launchConfiguration_copyPrivateIp :: Lens' LaunchConfiguration (Maybe Bool)

Copy Private IP during Launch Configuration.

UntagResource

untagResource_resourceArn :: Lens' UntagResource Text

Untag resource by ARN.

untagResource_tagKeys :: Lens' UntagResource [Text]

Untag resource by Keys.

DeleteSourceServer

deleteSourceServer_sourceServerID :: Lens' DeleteSourceServer Text

Request to delete Source Server from service by Server ID.

deleteSourceServerResponse_httpStatus :: Lens' DeleteSourceServerResponse Int

The response's http status code.

Types

CPU

cpu_modelName :: Lens' CPU (Maybe Text)

The source server's CPU model name.

cpu_cores :: Lens' CPU (Maybe Natural)

The number of CPU cores on the source server.

ChangeServerLifeCycleStateSourceServerLifecycle

changeServerLifeCycleStateSourceServerLifecycle_state :: Lens' ChangeServerLifeCycleStateSourceServerLifecycle ChangeServerLifeCycleStateSourceServerLifecycleState

The request to change the source server migration lifecycle state.

DataReplicationError

dataReplicationError_rawError :: Lens' DataReplicationError (Maybe Text)

Error in data replication.

dataReplicationError_error :: Lens' DataReplicationError (Maybe DataReplicationErrorString)

Error in data replication.

DataReplicationInfo

dataReplicationInfo_dataReplicationInitiation :: Lens' DataReplicationInfo (Maybe DataReplicationInitiation)

Request to query whether data replication has been initiated.

dataReplicationInfo_dataReplicationError :: Lens' DataReplicationInfo (Maybe DataReplicationError)

Error in obtaining data replication info.

dataReplicationInfo_lagDuration :: Lens' DataReplicationInfo (Maybe Text)

Request to query data replication lag durating.

dataReplicationInfo_dataReplicationState :: Lens' DataReplicationInfo (Maybe DataReplicationState)

Request to query the data replication state.

dataReplicationInfo_replicatedDisks :: Lens' DataReplicationInfo (Maybe [DataReplicationInfoReplicatedDisk])

Request to query disks replicated.

dataReplicationInfo_etaDateTime :: Lens' DataReplicationInfo (Maybe Text)

Request to query the time when data replication will be complete.

DataReplicationInfoReplicatedDisk

dataReplicationInfoReplicatedDisk_replicatedStorageBytes :: Lens' DataReplicationInfoReplicatedDisk (Maybe Natural)

Request to query amount of data replicated in bytes.

dataReplicationInfoReplicatedDisk_backloggedStorageBytes :: Lens' DataReplicationInfoReplicatedDisk (Maybe Natural)

Request to query data replication backlog size in bytes.

dataReplicationInfoReplicatedDisk_deviceName :: Lens' DataReplicationInfoReplicatedDisk (Maybe Text)

Request to query device name.

dataReplicationInfoReplicatedDisk_rescannedStorageBytes :: Lens' DataReplicationInfoReplicatedDisk (Maybe Natural)

Request to query amount of data rescanned in bytes.

dataReplicationInfoReplicatedDisk_totalStorageBytes :: Lens' DataReplicationInfoReplicatedDisk (Maybe Natural)

Request to query total amount of data replicated in bytes.

DataReplicationInitiation

dataReplicationInitiation_steps :: Lens' DataReplicationInitiation (Maybe [DataReplicationInitiationStep])

Request to query data initiation steps.

dataReplicationInitiation_nextAttemptDateTime :: Lens' DataReplicationInitiation (Maybe Text)

Request to query next data initiation date and time.

dataReplicationInitiation_startDateTime :: Lens' DataReplicationInitiation (Maybe Text)

Request to query data initiation start date and time.

DataReplicationInitiationStep

dataReplicationInitiationStep_status :: Lens' DataReplicationInitiationStep (Maybe DataReplicationInitiationStepStatus)

Request to query data initiation status.

dataReplicationInitiationStep_name :: Lens' DataReplicationInitiationStep (Maybe DataReplicationInitiationStepName)

Request to query data initiation step name.

DescribeJobsRequestFilters

describeJobsRequestFilters_fromDate :: Lens' DescribeJobsRequestFilters (Maybe Text)

Request to describe Job log filters by date.

describeJobsRequestFilters_toDate :: Lens' DescribeJobsRequestFilters (Maybe Text)

Request to describe Job log by last date.

describeJobsRequestFilters_jobIDs :: Lens' DescribeJobsRequestFilters (Maybe [Text])

Request to describe Job log filters by job ID.

DescribeSourceServersRequestFilters

describeSourceServersRequestFilters_sourceServerIDs :: Lens' DescribeSourceServersRequestFilters (Maybe [Text])

Request to filter Source Servers list by Source Server ID.

describeSourceServersRequestFilters_isArchived :: Lens' DescribeSourceServersRequestFilters (Maybe Bool)

Request to filter Source Servers list by archived.

Disk

disk_deviceName :: Lens' Disk (Maybe Text)

The disk or device name.

disk_bytes :: Lens' Disk (Maybe Natural)

The amount of storage on the disk in bytes.

IdentificationHints

identificationHints_hostname :: Lens' IdentificationHints (Maybe Text)

Hostname identification hint.

identificationHints_fqdn :: Lens' IdentificationHints (Maybe Text)

FQDN address identification hint.

identificationHints_awsInstanceID :: Lens' IdentificationHints (Maybe Text)

AWS Instance ID identification hint.

identificationHints_vmWareUuid :: Lens' IdentificationHints (Maybe Text)

vmWare UUID identification hint.

Job

job_initiatedBy :: Lens' Job (Maybe InitiatedBy)

Job initiated by field.

job_status :: Lens' Job (Maybe JobStatus)

Job status.

job_participatingServers :: Lens' Job (Maybe [ParticipatingServer])

Servers participating in a specific Job.

job_arn :: Lens' Job (Maybe Text)

the ARN of the specific Job.

job_creationDateTime :: Lens' Job (Maybe Text)

Job creation time.

job_type :: Lens' Job (Maybe JobType)

Job type.

job_endDateTime :: Lens' Job (Maybe Text)

Job end time.

job_tags :: Lens' Job (Maybe (HashMap Text Text))

Tags associated with spcific Job.

job_jobID :: Lens' Job Text

Job ID.

JobLog

jobLog_event :: Lens' JobLog (Maybe JobLogEvent)

Job log event.

jobLog_eventData :: Lens' JobLog (Maybe JobLogEventData)

Job event data

jobLog_logDateTime :: Lens' JobLog (Maybe Text)

Job log event date and time.

JobLogEventData

jobLogEventData_rawError :: Lens' JobLogEventData (Maybe Text)

Job error.

jobLogEventData_targetInstanceID :: Lens' JobLogEventData (Maybe Text)

Job Event Target instance ID.

jobLogEventData_sourceServerID :: Lens' JobLogEventData (Maybe Text)

Job Event Source Server ID.

jobLogEventData_conversionServerID :: Lens' JobLogEventData (Maybe Text)

Job Event conversion Server ID.

LaunchConfiguration

launchConfiguration_ec2LaunchTemplateID :: Lens' LaunchConfiguration (Maybe Text)

Configure EC2 lauch configuration template ID.

launchConfiguration_targetInstanceTypeRightSizingMethod :: Lens' LaunchConfiguration (Maybe TargetInstanceTypeRightSizingMethod)

Configure launch configuration Target instance type right sizing method.

launchConfiguration_launchDisposition :: Lens' LaunchConfiguration (Maybe LaunchDisposition)

Configure launch dispostion for launch configuration.

launchConfiguration_copyTags :: Lens' LaunchConfiguration (Maybe Bool)

Copy Tags during Launch Configuration.

launchConfiguration_name :: Lens' LaunchConfiguration (Maybe Text)

Configure launch configuration name.

launchConfiguration_sourceServerID :: Lens' LaunchConfiguration (Maybe Text)

Configure launch configuration Source Server ID.

launchConfiguration_licensing :: Lens' LaunchConfiguration (Maybe Licensing)

Configure launch configuration OS licensing.

launchConfiguration_copyPrivateIp :: Lens' LaunchConfiguration (Maybe Bool)

Copy Private IP during Launch Configuration.

LaunchedInstance

launchedInstance_jobID :: Lens' LaunchedInstance (Maybe Text)

Configure launced instance Job ID.

launchedInstance_ec2InstanceID :: Lens' LaunchedInstance (Maybe Text)

Configure launced instance EC2 ID.

launchedInstance_firstBoot :: Lens' LaunchedInstance (Maybe FirstBoot)

Configure launced instance first boot.

Licensing

licensing_osByol :: Lens' Licensing (Maybe Bool)

Configure BYOL OS licensing.

LifeCycle

lifeCycle_lastTest :: Lens' LifeCycle (Maybe LifeCycleLastTest)

Lifecycle last Test.

lifeCycle_state :: Lens' LifeCycle (Maybe LifeCycleState)

Lifecycle state.

lifeCycle_elapsedReplicationDuration :: Lens' LifeCycle (Maybe Text)

Lifecycle elapsed time and duration.

lifeCycle_lastSeenByServiceDateTime :: Lens' LifeCycle (Maybe Text)

Lifecycle last seen date and time.

lifeCycle_addedToServiceDateTime :: Lens' LifeCycle (Maybe Text)

Lifecycle added to service data and time.

lifeCycle_lastCutover :: Lens' LifeCycle (Maybe LifeCycleLastCutover)

Lifecycle last Cutover.

lifeCycle_firstByteDateTime :: Lens' LifeCycle (Maybe Text)

Lifecycle replication initiation date and time.

LifeCycleLastCutover

lifeCycleLastCutover_initiated :: Lens' LifeCycleLastCutover (Maybe LifeCycleLastCutoverInitiated)

Lifecycle last Cutover initiated.

lifeCycleLastCutover_reverted :: Lens' LifeCycleLastCutover (Maybe LifeCycleLastCutoverReverted)

Lifecycle last Cutover reverted.

lifeCycleLastCutover_finalized :: Lens' LifeCycleLastCutover (Maybe LifeCycleLastCutoverFinalized)

Lifecycle Cutover finalized date and time.

LifeCycleLastCutoverFinalized

lifeCycleLastCutoverFinalized_apiCallDateTime :: Lens' LifeCycleLastCutoverFinalized (Maybe Text)

Lifecycle Cutover finalized date and time.

LifeCycleLastCutoverInitiated

lifeCycleLastCutoverInitiated_jobID :: Lens' LifeCycleLastCutoverInitiated (Maybe Text)

Lifecycle last Cutover initiated by Job ID.

lifeCycleLastCutoverInitiated_apiCallDateTime :: Lens' LifeCycleLastCutoverInitiated (Maybe Text)

LifeCycleLastCutoverReverted

lifeCycleLastCutoverReverted_apiCallDateTime :: Lens' LifeCycleLastCutoverReverted (Maybe Text)

Lifecycle last Cutover reverted API call date time.

LifeCycleLastTest

lifeCycleLastTest_initiated :: Lens' LifeCycleLastTest (Maybe LifeCycleLastTestInitiated)

Lifecycle last Test initiated.

lifeCycleLastTest_reverted :: Lens' LifeCycleLastTest (Maybe LifeCycleLastTestReverted)

Lifecycle last Test reverted.

lifeCycleLastTest_finalized :: Lens' LifeCycleLastTest (Maybe LifeCycleLastTestFinalized)

Lifecycle last Test finlized.

LifeCycleLastTestFinalized

lifeCycleLastTestFinalized_apiCallDateTime :: Lens' LifeCycleLastTestFinalized (Maybe Text)

Lifecycle Test failed API call date and time.

LifeCycleLastTestInitiated

lifeCycleLastTestInitiated_jobID :: Lens' LifeCycleLastTestInitiated (Maybe Text)

Lifecycle last Test initiated Job ID.

lifeCycleLastTestInitiated_apiCallDateTime :: Lens' LifeCycleLastTestInitiated (Maybe Text)

Lifecycle last Test initiated API call date and time.

LifeCycleLastTestReverted

lifeCycleLastTestReverted_apiCallDateTime :: Lens' LifeCycleLastTestReverted (Maybe Text)

Lifecycle last Test reverted API call date and time.

NetworkInterface

networkInterface_ips :: Lens' NetworkInterface (Maybe [Text])

Network interface IPs.

networkInterface_macAddress :: Lens' NetworkInterface (Maybe Text)

Network interface Mac address.

networkInterface_isPrimary :: Lens' NetworkInterface (Maybe Bool)

Network interface primary IP.

OS

os_fullString :: Lens' OS (Maybe Text)

OS full string.

ParticipatingServer

participatingServer_launchStatus :: Lens' ParticipatingServer (Maybe LaunchStatus)

Participating server launch status.

participatingServer_sourceServerID :: Lens' ParticipatingServer (Maybe Text)

Participating server Source Server ID.

ReplicationConfiguration

replicationConfiguration_createPublicIP :: Lens' ReplicationConfiguration (Maybe Bool)

Replication Configuration create Public IP.

replicationConfiguration_stagingAreaTags :: Lens' ReplicationConfiguration (Maybe (HashMap Text Text))

Replication Configuration Staging Area tags.

replicationConfiguration_stagingAreaSubnetId :: Lens' ReplicationConfiguration (Maybe Text)

Replication Configuration Staging Area subnet ID.

replicationConfiguration_replicationServerInstanceType :: Lens' ReplicationConfiguration (Maybe Text)

Replication Configuration Replication Server instance type.

replicationConfiguration_ebsEncryption :: Lens' ReplicationConfiguration (Maybe ReplicationConfigurationEbsEncryption)

Replication Configuration EBS encryption.

replicationConfiguration_associateDefaultSecurityGroup :: Lens' ReplicationConfiguration (Maybe Bool)

Replication Configuration associate default Application Migration Service Security Group.

replicationConfiguration_replicationServersSecurityGroupsIDs :: Lens' ReplicationConfiguration (Maybe [Text])

Replication Configuration Replication Server Security Group IDs.

replicationConfiguration_ebsEncryptionKeyArn :: Lens' ReplicationConfiguration (Maybe Text)

Replication Configuration EBS encryption key ARN.

replicationConfiguration_name :: Lens' ReplicationConfiguration (Maybe Text)

Replication Configuration name.

replicationConfiguration_defaultLargeStagingDiskType :: Lens' ReplicationConfiguration (Maybe ReplicationConfigurationDefaultLargeStagingDiskType)

Replication Configuration use default large Staging Disks.

replicationConfiguration_bandwidthThrottling :: Lens' ReplicationConfiguration (Maybe Natural)

Replication Configuration set bandwidth throttling.

replicationConfiguration_dataPlaneRouting :: Lens' ReplicationConfiguration (Maybe ReplicationConfigurationDataPlaneRouting)

Replication Configuration data plane routing.

replicationConfiguration_sourceServerID :: Lens' ReplicationConfiguration (Maybe Text)

Replication Configuration Source Server ID.

replicationConfiguration_replicatedDisks :: Lens' ReplicationConfiguration (Maybe [ReplicationConfigurationReplicatedDisk])

Replication Configuration replicated disks.

replicationConfiguration_useDedicatedReplicationServer :: Lens' ReplicationConfiguration (Maybe Bool)

Replication Configuration use Dedicated Replication Server.

ReplicationConfigurationReplicatedDisk

replicationConfigurationReplicatedDisk_stagingDiskType :: Lens' ReplicationConfigurationReplicatedDisk (Maybe ReplicationConfigurationReplicatedDiskStagingDiskType)

Replication Configuration replicated disk staging disk type.

replicationConfigurationReplicatedDisk_iops :: Lens' ReplicationConfigurationReplicatedDisk (Maybe Natural)

Replication Configuration replicated disk IOPs.

replicationConfigurationReplicatedDisk_isBootDisk :: Lens' ReplicationConfigurationReplicatedDisk (Maybe Bool)

Replication Configuration replicated disk boot disk.

replicationConfigurationReplicatedDisk_deviceName :: Lens' ReplicationConfigurationReplicatedDisk (Maybe Text)

Replication Configuration replicated disk device name.

ReplicationConfigurationTemplate

replicationConfigurationTemplate_createPublicIP :: Lens' ReplicationConfigurationTemplate (Maybe Bool)

Replication Configuration template create Public IP.

replicationConfigurationTemplate_stagingAreaTags :: Lens' ReplicationConfigurationTemplate (Maybe (HashMap Text Text))

Replication Configuration template Staging Area Tags.

replicationConfigurationTemplate_arn :: Lens' ReplicationConfigurationTemplate (Maybe Text)

Replication Configuration template ARN.

replicationConfigurationTemplate_stagingAreaSubnetId :: Lens' ReplicationConfigurationTemplate (Maybe Text)

Replication Configuration template Staging Area subnet ID.

replicationConfigurationTemplate_replicationServerInstanceType :: Lens' ReplicationConfigurationTemplate (Maybe Text)

Replication Configuration template server instance type.

replicationConfigurationTemplate_ebsEncryption :: Lens' ReplicationConfigurationTemplate (Maybe ReplicationConfigurationEbsEncryption)

Replication Configuration template EBS encryption.

replicationConfigurationTemplate_associateDefaultSecurityGroup :: Lens' ReplicationConfigurationTemplate (Maybe Bool)

Replication Configuration template associate default Application Migration Service Security group.

replicationConfigurationTemplate_replicationServersSecurityGroupsIDs :: Lens' ReplicationConfigurationTemplate (Maybe [Text])

Replication Configuration template server Security Groups IDs.

replicationConfigurationTemplate_ebsEncryptionKeyArn :: Lens' ReplicationConfigurationTemplate (Maybe Text)

Replication Configuration template EBS encryption key ARN.

replicationConfigurationTemplate_defaultLargeStagingDiskType :: Lens' ReplicationConfigurationTemplate (Maybe ReplicationConfigurationDefaultLargeStagingDiskType)

Replication Configuration template use dedault large Staging Disk type.

replicationConfigurationTemplate_bandwidthThrottling :: Lens' ReplicationConfigurationTemplate (Maybe Natural)

Replication Configuration template bandwidth throtting.

replicationConfigurationTemplate_dataPlaneRouting :: Lens' ReplicationConfigurationTemplate (Maybe ReplicationConfigurationDataPlaneRouting)

Replication Configuration template data plane routing.

replicationConfigurationTemplate_useDedicatedReplicationServer :: Lens' ReplicationConfigurationTemplate (Maybe Bool)

Replication Configuration template use Dedicated Replication Server.

replicationConfigurationTemplate_tags :: Lens' ReplicationConfigurationTemplate (Maybe (HashMap Text Text))

Replication Configuration template Tags.

replicationConfigurationTemplate_replicationConfigurationTemplateID :: Lens' ReplicationConfigurationTemplate Text

Replication Configuration template template ID.

SourceProperties

sourceProperties_identificationHints :: Lens' SourceProperties (Maybe IdentificationHints)

Source server identification hints.

sourceProperties_networkInterfaces :: Lens' SourceProperties (Maybe [NetworkInterface])

Source server network interfaces.

sourceProperties_lastUpdatedDateTime :: Lens' SourceProperties (Maybe Text)

Source server last update date and time.

sourceProperties_recommendedInstanceType :: Lens' SourceProperties (Maybe Text)

Source server recommended instance type.

sourceProperties_os :: Lens' SourceProperties (Maybe OS)

Source server OS.

sourceProperties_ramBytes :: Lens' SourceProperties (Maybe Natural)

Source server RAM in bytes.

sourceProperties_cpus :: Lens' SourceProperties (Maybe [CPU])

Source Server CPUs.

sourceProperties_disks :: Lens' SourceProperties (Maybe [Disk])

Source Server disks.

SourceServer

sourceServer_sourceProperties :: Lens' SourceServer (Maybe SourceProperties)

Source server properties.

sourceServer_arn :: Lens' SourceServer (Maybe Text)

Source server ARN.

sourceServer_launchedInstance :: Lens' SourceServer (Maybe LaunchedInstance)

Source server launched instance.

sourceServer_lifeCycle :: Lens' SourceServer (Maybe LifeCycle)

Source server lifecycle state.

sourceServer_isArchived :: Lens' SourceServer (Maybe Bool)

Source server archived status.

sourceServer_dataReplicationInfo :: Lens' SourceServer (Maybe DataReplicationInfo)

Source server data replication info.

sourceServer_sourceServerID :: Lens' SourceServer (Maybe Text)

Source server ID.

sourceServer_tags :: Lens' SourceServer (Maybe (HashMap Text Text))

Source server Tags.