Amazonka.MGN.Types.DataReplicationInfoReplicatedDisk

Description

Documentation

data DataReplicationInfoReplicatedDisk

Request to query disks replicated.

See: newDataReplicationInfoReplicatedDisk smart constructor.

Constructors

DataReplicationInfoReplicatedDisk'

Fields

replicatedStorageBytes :: Maybe Natural

Request to query amount of data replicated in bytes.

backloggedStorageBytes :: Maybe Natural

Request to query data replication backlog size in bytes.

deviceName :: Maybe Text

Request to query device name.

rescannedStorageBytes :: Maybe Natural

Request to query amount of data rescanned in bytes.

totalStorageBytes :: Maybe Natural

Request to query total amount of data replicated in bytes.

newDataReplicationInfoReplicatedDisk :: DataReplicationInfoReplicatedDisk

Create a value of DataReplicationInfoReplicatedDisk with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:replicatedStorageBytes:DataReplicationInfoReplicatedDisk', dataReplicationInfoReplicatedDisk_replicatedStorageBytes - Request to query amount of data replicated in bytes.

$sel:backloggedStorageBytes:DataReplicationInfoReplicatedDisk', dataReplicationInfoReplicatedDisk_backloggedStorageBytes - Request to query data replication backlog size in bytes.

$sel:deviceName:DataReplicationInfoReplicatedDisk', dataReplicationInfoReplicatedDisk_deviceName - Request to query device name.

$sel:rescannedStorageBytes:DataReplicationInfoReplicatedDisk', dataReplicationInfoReplicatedDisk_rescannedStorageBytes - Request to query amount of data rescanned in bytes.

$sel:totalStorageBytes:DataReplicationInfoReplicatedDisk', dataReplicationInfoReplicatedDisk_totalStorageBytes - Request to query total amount of data replicated in bytes.

dataReplicationInfoReplicatedDisk_replicatedStorageBytes :: Lens' DataReplicationInfoReplicatedDisk (Maybe Natural)

Request to query amount of data replicated in bytes.

dataReplicationInfoReplicatedDisk_backloggedStorageBytes :: Lens' DataReplicationInfoReplicatedDisk (Maybe Natural)

Request to query data replication backlog size in bytes.

dataReplicationInfoReplicatedDisk_deviceName :: Lens' DataReplicationInfoReplicatedDisk (Maybe Text)

Request to query device name.

dataReplicationInfoReplicatedDisk_rescannedStorageBytes :: Lens' DataReplicationInfoReplicatedDisk (Maybe Natural)

Request to query amount of data rescanned in bytes.

dataReplicationInfoReplicatedDisk_totalStorageBytes :: Lens' DataReplicationInfoReplicatedDisk (Maybe Natural)

Request to query total amount of data replicated in bytes.