Amazonka.MGN.Types.ReplicationConfigurationTemplate

Description

Documentation

data ReplicationConfigurationTemplate

See: newReplicationConfigurationTemplate smart constructor.

Constructors

ReplicationConfigurationTemplate'

Fields

createPublicIP :: Maybe Bool

Replication Configuration template create Public IP.

stagingAreaTags :: Maybe (Sensitive (HashMap Text Text))

Replication Configuration template Staging Area Tags.

arn :: Maybe Text

Replication Configuration template ARN.

stagingAreaSubnetId :: Maybe Text

Replication Configuration template Staging Area subnet ID.

replicationServerInstanceType :: Maybe Text

Replication Configuration template server instance type.

ebsEncryption :: Maybe ReplicationConfigurationEbsEncryption

Replication Configuration template EBS encryption.

associateDefaultSecurityGroup :: Maybe Bool

Replication Configuration template associate default Application Migration Service Security group.

replicationServersSecurityGroupsIDs :: Maybe [Text]

Replication Configuration template server Security Groups IDs.

ebsEncryptionKeyArn :: Maybe Text

Replication Configuration template EBS encryption key ARN.

defaultLargeStagingDiskType :: Maybe ReplicationConfigurationDefaultLargeStagingDiskType

Replication Configuration template use dedault large Staging Disk type.

bandwidthThrottling :: Maybe Natural

Replication Configuration template bandwidth throtting.

dataPlaneRouting :: Maybe ReplicationConfigurationDataPlaneRouting

Replication Configuration template data plane routing.

useDedicatedReplicationServer :: Maybe Bool

Replication Configuration template use Dedicated Replication Server.

tags :: Maybe (Sensitive (HashMap Text Text))

Replication Configuration template Tags.

replicationConfigurationTemplateID :: Text

Replication Configuration template template ID.

newReplicationConfigurationTemplate

Arguments

:: Text
$sel:replicationConfigurationTemplateID:ReplicationConfigurationTemplate'

-> ReplicationConfigurationTemplate

Create a value of ReplicationConfigurationTemplate with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:createPublicIP:ReplicationConfigurationTemplate', replicationConfigurationTemplate_createPublicIP - Replication Configuration template create Public IP.

$sel:stagingAreaTags:ReplicationConfigurationTemplate', replicationConfigurationTemplate_stagingAreaTags - Replication Configuration template Staging Area Tags.

$sel:arn:ReplicationConfigurationTemplate', replicationConfigurationTemplate_arn - Replication Configuration template ARN.

$sel:stagingAreaSubnetId:ReplicationConfigurationTemplate', replicationConfigurationTemplate_stagingAreaSubnetId - Replication Configuration template Staging Area subnet ID.

$sel:replicationServerInstanceType:ReplicationConfigurationTemplate', replicationConfigurationTemplate_replicationServerInstanceType - Replication Configuration template server instance type.

$sel:ebsEncryption:ReplicationConfigurationTemplate', replicationConfigurationTemplate_ebsEncryption - Replication Configuration template EBS encryption.

$sel:associateDefaultSecurityGroup:ReplicationConfigurationTemplate', replicationConfigurationTemplate_associateDefaultSecurityGroup - Replication Configuration template associate default Application Migration Service Security group.

$sel:replicationServersSecurityGroupsIDs:ReplicationConfigurationTemplate', replicationConfigurationTemplate_replicationServersSecurityGroupsIDs - Replication Configuration template server Security Groups IDs.

$sel:ebsEncryptionKeyArn:ReplicationConfigurationTemplate', replicationConfigurationTemplate_ebsEncryptionKeyArn - Replication Configuration template EBS encryption key ARN.

$sel:defaultLargeStagingDiskType:ReplicationConfigurationTemplate', replicationConfigurationTemplate_defaultLargeStagingDiskType - Replication Configuration template use dedault large Staging Disk type.

$sel:bandwidthThrottling:ReplicationConfigurationTemplate', replicationConfigurationTemplate_bandwidthThrottling - Replication Configuration template bandwidth throtting.

$sel:dataPlaneRouting:ReplicationConfigurationTemplate', replicationConfigurationTemplate_dataPlaneRouting - Replication Configuration template data plane routing.

$sel:useDedicatedReplicationServer:ReplicationConfigurationTemplate', replicationConfigurationTemplate_useDedicatedReplicationServer - Replication Configuration template use Dedicated Replication Server.

$sel:tags:ReplicationConfigurationTemplate', replicationConfigurationTemplate_tags - Replication Configuration template Tags.

$sel:replicationConfigurationTemplateID:ReplicationConfigurationTemplate', replicationConfigurationTemplate_replicationConfigurationTemplateID - Replication Configuration template template ID.

replicationConfigurationTemplate_createPublicIP :: Lens' ReplicationConfigurationTemplate (Maybe Bool)

Replication Configuration template create Public IP.

replicationConfigurationTemplate_stagingAreaTags :: Lens' ReplicationConfigurationTemplate (Maybe (HashMap Text Text))

Replication Configuration template Staging Area Tags.

replicationConfigurationTemplate_arn :: Lens' ReplicationConfigurationTemplate (Maybe Text)

Replication Configuration template ARN.

replicationConfigurationTemplate_stagingAreaSubnetId :: Lens' ReplicationConfigurationTemplate (Maybe Text)

Replication Configuration template Staging Area subnet ID.

replicationConfigurationTemplate_replicationServerInstanceType :: Lens' ReplicationConfigurationTemplate (Maybe Text)

Replication Configuration template server instance type.

replicationConfigurationTemplate_ebsEncryption :: Lens' ReplicationConfigurationTemplate (Maybe ReplicationConfigurationEbsEncryption)

Replication Configuration template EBS encryption.

replicationConfigurationTemplate_associateDefaultSecurityGroup :: Lens' ReplicationConfigurationTemplate (Maybe Bool)

Replication Configuration template associate default Application Migration Service Security group.

replicationConfigurationTemplate_replicationServersSecurityGroupsIDs :: Lens' ReplicationConfigurationTemplate (Maybe [Text])

Replication Configuration template server Security Groups IDs.

replicationConfigurationTemplate_ebsEncryptionKeyArn :: Lens' ReplicationConfigurationTemplate (Maybe Text)

Replication Configuration template EBS encryption key ARN.

replicationConfigurationTemplate_defaultLargeStagingDiskType :: Lens' ReplicationConfigurationTemplate (Maybe ReplicationConfigurationDefaultLargeStagingDiskType)

Replication Configuration template use dedault large Staging Disk type.

replicationConfigurationTemplate_bandwidthThrottling :: Lens' ReplicationConfigurationTemplate (Maybe Natural)

Replication Configuration template bandwidth throtting.

replicationConfigurationTemplate_dataPlaneRouting :: Lens' ReplicationConfigurationTemplate (Maybe ReplicationConfigurationDataPlaneRouting)

Replication Configuration template data plane routing.

replicationConfigurationTemplate_useDedicatedReplicationServer :: Lens' ReplicationConfigurationTemplate (Maybe Bool)

Replication Configuration template use Dedicated Replication Server.

replicationConfigurationTemplate_tags :: Lens' ReplicationConfigurationTemplate (Maybe (HashMap Text Text))

Replication Configuration template Tags.

replicationConfigurationTemplate_replicationConfigurationTemplateID :: Lens' ReplicationConfigurationTemplate Text

Replication Configuration template template ID.