Amazonka.MGN.UpdateLaunchConfiguration

Description

Updates multiple LaunchConfigurations by Source Server ID.

Creating a Request

data UpdateLaunchConfiguration

See: newUpdateLaunchConfiguration smart constructor.

Constructors

UpdateLaunchConfiguration'

Fields

targetInstanceTypeRightSizingMethod :: Maybe TargetInstanceTypeRightSizingMethod

Update Launch configuration Target instance right sizing request.

launchDisposition :: Maybe LaunchDisposition

Update Launch configuration launch disposition request.

copyTags :: Maybe Bool

Update Launch configuration copy Tags request.

name :: Maybe Text

Update Launch configuration name request.

licensing :: Maybe Licensing

Update Launch configuration licensing request.

copyPrivateIp :: Maybe Bool

Update Launch configuration copy Private IP request.

sourceServerID :: Text

Update Launch configuration by Source Server ID request.

newUpdateLaunchConfiguration

Arguments

:: Text
$sel:sourceServerID:UpdateLaunchConfiguration'

-> UpdateLaunchConfiguration

Create a value of UpdateLaunchConfiguration with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:targetInstanceTypeRightSizingMethod:UpdateLaunchConfiguration', updateLaunchConfiguration_targetInstanceTypeRightSizingMethod - Update Launch configuration Target instance right sizing request.

$sel:launchDisposition:UpdateLaunchConfiguration', updateLaunchConfiguration_launchDisposition - Update Launch configuration launch disposition request.

$sel:copyTags:UpdateLaunchConfiguration', updateLaunchConfiguration_copyTags - Update Launch configuration copy Tags request.

$sel:name:UpdateLaunchConfiguration', updateLaunchConfiguration_name - Update Launch configuration name request.

$sel:licensing:UpdateLaunchConfiguration', updateLaunchConfiguration_licensing - Update Launch configuration licensing request.

$sel:copyPrivateIp:UpdateLaunchConfiguration', updateLaunchConfiguration_copyPrivateIp - Update Launch configuration copy Private IP request.

$sel:sourceServerID:UpdateLaunchConfiguration', updateLaunchConfiguration_sourceServerID - Update Launch configuration by Source Server ID request.

Request Lenses

updateLaunchConfiguration_targetInstanceTypeRightSizingMethod :: Lens' UpdateLaunchConfiguration (Maybe TargetInstanceTypeRightSizingMethod)

Update Launch configuration Target instance right sizing request.

updateLaunchConfiguration_launchDisposition :: Lens' UpdateLaunchConfiguration (Maybe LaunchDisposition)

Update Launch configuration launch disposition request.

updateLaunchConfiguration_copyTags :: Lens' UpdateLaunchConfiguration (Maybe Bool)

Update Launch configuration copy Tags request.

updateLaunchConfiguration_name :: Lens' UpdateLaunchConfiguration (Maybe Text)

Update Launch configuration name request.

updateLaunchConfiguration_licensing :: Lens' UpdateLaunchConfiguration (Maybe Licensing)

Update Launch configuration licensing request.

updateLaunchConfiguration_copyPrivateIp :: Lens' UpdateLaunchConfiguration (Maybe Bool)

Update Launch configuration copy Private IP request.

updateLaunchConfiguration_sourceServerID :: Lens' UpdateLaunchConfiguration Text

Update Launch configuration by Source Server ID request.

Destructuring the Response

data LaunchConfiguration

See: newLaunchConfiguration smart constructor.

Constructors

LaunchConfiguration'

Fields

ec2LaunchTemplateID :: Maybe Text

Configure EC2 lauch configuration template ID.

targetInstanceTypeRightSizingMethod :: Maybe TargetInstanceTypeRightSizingMethod

Configure launch configuration Target instance type right sizing method.

launchDisposition :: Maybe LaunchDisposition

Configure launch dispostion for launch configuration.

copyTags :: Maybe Bool

Copy Tags during Launch Configuration.

name :: Maybe Text

Configure launch configuration name.

sourceServerID :: Maybe Text

Configure launch configuration Source Server ID.

licensing :: Maybe Licensing

Configure launch configuration OS licensing.

copyPrivateIp :: Maybe Bool

Copy Private IP during Launch Configuration.

newLaunchConfiguration :: LaunchConfiguration

Create a value of LaunchConfiguration with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:ec2LaunchTemplateID:LaunchConfiguration', launchConfiguration_ec2LaunchTemplateID - Configure EC2 lauch configuration template ID.

$sel:targetInstanceTypeRightSizingMethod:LaunchConfiguration', launchConfiguration_targetInstanceTypeRightSizingMethod - Configure launch configuration Target instance type right sizing method.

$sel:launchDisposition:LaunchConfiguration', launchConfiguration_launchDisposition - Configure launch dispostion for launch configuration.

$sel:copyTags:LaunchConfiguration', launchConfiguration_copyTags - Copy Tags during Launch Configuration.

$sel:name:LaunchConfiguration', launchConfiguration_name - Configure launch configuration name.

$sel:sourceServerID:LaunchConfiguration', launchConfiguration_sourceServerID - Configure launch configuration Source Server ID.

$sel:licensing:LaunchConfiguration', launchConfiguration_licensing - Configure launch configuration OS licensing.

$sel:copyPrivateIp:LaunchConfiguration', launchConfiguration_copyPrivateIp - Copy Private IP during Launch Configuration.

Response Lenses

launchConfiguration_ec2LaunchTemplateID :: Lens' LaunchConfiguration (Maybe Text)

Configure EC2 lauch configuration template ID.

launchConfiguration_targetInstanceTypeRightSizingMethod :: Lens' LaunchConfiguration (Maybe TargetInstanceTypeRightSizingMethod)

Configure launch configuration Target instance type right sizing method.

launchConfiguration_launchDisposition :: Lens' LaunchConfiguration (Maybe LaunchDisposition)

Configure launch dispostion for launch configuration.

launchConfiguration_copyTags :: Lens' LaunchConfiguration (Maybe Bool)

Copy Tags during Launch Configuration.

launchConfiguration_name :: Lens' LaunchConfiguration (Maybe Text)

Configure launch configuration name.

launchConfiguration_sourceServerID :: Lens' LaunchConfiguration (Maybe Text)

Configure launch configuration Source Server ID.

launchConfiguration_licensing :: Lens' LaunchConfiguration (Maybe Licensing)

Configure launch configuration OS licensing.

launchConfiguration_copyPrivateIp :: Lens' LaunchConfiguration (Maybe Bool)

Copy Private IP during Launch Configuration.