Amazonka.MGN.UpdateReplicationConfigurationTemplate

Description

Updates multiple ReplicationConfigurationTemplates by ID.

Creating a Request

data UpdateReplicationConfigurationTemplate

See: newUpdateReplicationConfigurationTemplate smart constructor.

Constructors

UpdateReplicationConfigurationTemplate'

Fields

createPublicIP :: Maybe Bool

Update replication configuration template create Public IP request.

stagingAreaTags :: Maybe (Sensitive (HashMap Text Text))

Update replication configuration template Staging Area Tags request.

arn :: Maybe Text

Update replication configuration template ARN request.

stagingAreaSubnetId :: Maybe Text

Update replication configuration template Staging Area subnet ID request.

replicationServerInstanceType :: Maybe Text

Update replication configuration template Replication Server instance type request.

ebsEncryption :: Maybe ReplicationConfigurationEbsEncryption

Update replication configuration template EBS encryption request.

associateDefaultSecurityGroup :: Maybe Bool

Update replication configuration template associate default Application Migration Service Security group request.

replicationServersSecurityGroupsIDs :: Maybe [Text]

Update replication configuration template Replication Server Security groups IDs request.

ebsEncryptionKeyArn :: Maybe Text

Update replication configuration template EBS encryption key ARN request.

defaultLargeStagingDiskType :: Maybe ReplicationConfigurationDefaultLargeStagingDiskType

Update replication configuration template use default large Staging Disk type request.

bandwidthThrottling :: Maybe Natural

Update replication configuration template bandwidth throttling request.

dataPlaneRouting :: Maybe ReplicationConfigurationDataPlaneRouting

Update replication configuration template data plane routing request.

useDedicatedReplicationServer :: Maybe Bool

Update replication configuration template use dedicated Replication Server request.

replicationConfigurationTemplateID :: Text

Update replication configuration template template ID request.

newUpdateReplicationConfigurationTemplate

Arguments

:: Text
$sel:replicationConfigurationTemplateID:UpdateReplicationConfigurationTemplate'

-> UpdateReplicationConfigurationTemplate

Create a value of UpdateReplicationConfigurationTemplate with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:createPublicIP:UpdateReplicationConfigurationTemplate', updateReplicationConfigurationTemplate_createPublicIP - Update replication configuration template create Public IP request.

$sel:stagingAreaTags:UpdateReplicationConfigurationTemplate', updateReplicationConfigurationTemplate_stagingAreaTags - Update replication configuration template Staging Area Tags request.

$sel:arn:UpdateReplicationConfigurationTemplate', updateReplicationConfigurationTemplate_arn - Update replication configuration template ARN request.

$sel:stagingAreaSubnetId:UpdateReplicationConfigurationTemplate', updateReplicationConfigurationTemplate_stagingAreaSubnetId - Update replication configuration template Staging Area subnet ID request.

$sel:replicationServerInstanceType:UpdateReplicationConfigurationTemplate', updateReplicationConfigurationTemplate_replicationServerInstanceType - Update replication configuration template Replication Server instance type request.

$sel:ebsEncryption:UpdateReplicationConfigurationTemplate', updateReplicationConfigurationTemplate_ebsEncryption - Update replication configuration template EBS encryption request.

$sel:associateDefaultSecurityGroup:UpdateReplicationConfigurationTemplate', updateReplicationConfigurationTemplate_associateDefaultSecurityGroup - Update replication configuration template associate default Application Migration Service Security group request.

$sel:replicationServersSecurityGroupsIDs:UpdateReplicationConfigurationTemplate', updateReplicationConfigurationTemplate_replicationServersSecurityGroupsIDs - Update replication configuration template Replication Server Security groups IDs request.

$sel:ebsEncryptionKeyArn:UpdateReplicationConfigurationTemplate', updateReplicationConfigurationTemplate_ebsEncryptionKeyArn - Update replication configuration template EBS encryption key ARN request.

$sel:defaultLargeStagingDiskType:UpdateReplicationConfigurationTemplate', updateReplicationConfigurationTemplate_defaultLargeStagingDiskType - Update replication configuration template use default large Staging Disk type request.

$sel:bandwidthThrottling:UpdateReplicationConfigurationTemplate', updateReplicationConfigurationTemplate_bandwidthThrottling - Update replication configuration template bandwidth throttling request.

$sel:dataPlaneRouting:UpdateReplicationConfigurationTemplate', updateReplicationConfigurationTemplate_dataPlaneRouting - Update replication configuration template data plane routing request.

$sel:useDedicatedReplicationServer:UpdateReplicationConfigurationTemplate', updateReplicationConfigurationTemplate_useDedicatedReplicationServer - Update replication configuration template use dedicated Replication Server request.

$sel:replicationConfigurationTemplateID:UpdateReplicationConfigurationTemplate', updateReplicationConfigurationTemplate_replicationConfigurationTemplateID - Update replication configuration template template ID request.

Request Lenses

updateReplicationConfigurationTemplate_createPublicIP :: Lens' UpdateReplicationConfigurationTemplate (Maybe Bool)

Update replication configuration template create Public IP request.

updateReplicationConfigurationTemplate_stagingAreaTags :: Lens' UpdateReplicationConfigurationTemplate (Maybe (HashMap Text Text))

Update replication configuration template Staging Area Tags request.

updateReplicationConfigurationTemplate_arn :: Lens' UpdateReplicationConfigurationTemplate (Maybe Text)

Update replication configuration template ARN request.

updateReplicationConfigurationTemplate_stagingAreaSubnetId :: Lens' UpdateReplicationConfigurationTemplate (Maybe Text)

Update replication configuration template Staging Area subnet ID request.

updateReplicationConfigurationTemplate_replicationServerInstanceType :: Lens' UpdateReplicationConfigurationTemplate (Maybe Text)

Update replication configuration template Replication Server instance type request.

updateReplicationConfigurationTemplate_ebsEncryption :: Lens' UpdateReplicationConfigurationTemplate (Maybe ReplicationConfigurationEbsEncryption)

Update replication configuration template EBS encryption request.

updateReplicationConfigurationTemplate_associateDefaultSecurityGroup :: Lens' UpdateReplicationConfigurationTemplate (Maybe Bool)

Update replication configuration template associate default Application Migration Service Security group request.

updateReplicationConfigurationTemplate_replicationServersSecurityGroupsIDs :: Lens' UpdateReplicationConfigurationTemplate (Maybe [Text])

Update replication configuration template Replication Server Security groups IDs request.

updateReplicationConfigurationTemplate_ebsEncryptionKeyArn :: Lens' UpdateReplicationConfigurationTemplate (Maybe Text)

Update replication configuration template EBS encryption key ARN request.

updateReplicationConfigurationTemplate_defaultLargeStagingDiskType :: Lens' UpdateReplicationConfigurationTemplate (Maybe ReplicationConfigurationDefaultLargeStagingDiskType)

Update replication configuration template use default large Staging Disk type request.

updateReplicationConfigurationTemplate_bandwidthThrottling :: Lens' UpdateReplicationConfigurationTemplate (Maybe Natural)

Update replication configuration template bandwidth throttling request.

updateReplicationConfigurationTemplate_dataPlaneRouting :: Lens' UpdateReplicationConfigurationTemplate (Maybe ReplicationConfigurationDataPlaneRouting)

Update replication configuration template data plane routing request.

updateReplicationConfigurationTemplate_useDedicatedReplicationServer :: Lens' UpdateReplicationConfigurationTemplate (Maybe Bool)

Update replication configuration template use dedicated Replication Server request.

updateReplicationConfigurationTemplate_replicationConfigurationTemplateID :: Lens' UpdateReplicationConfigurationTemplate Text

Update replication configuration template template ID request.

Destructuring the Response

data ReplicationConfigurationTemplate

See: newReplicationConfigurationTemplate smart constructor.

Constructors

ReplicationConfigurationTemplate'

Fields

createPublicIP :: Maybe Bool

Replication Configuration template create Public IP.

stagingAreaTags :: Maybe (Sensitive (HashMap Text Text))

Replication Configuration template Staging Area Tags.

arn :: Maybe Text

Replication Configuration template ARN.

stagingAreaSubnetId :: Maybe Text

Replication Configuration template Staging Area subnet ID.

replicationServerInstanceType :: Maybe Text

Replication Configuration template server instance type.

ebsEncryption :: Maybe ReplicationConfigurationEbsEncryption

Replication Configuration template EBS encryption.

associateDefaultSecurityGroup :: Maybe Bool

Replication Configuration template associate default Application Migration Service Security group.

replicationServersSecurityGroupsIDs :: Maybe [Text]

Replication Configuration template server Security Groups IDs.

ebsEncryptionKeyArn :: Maybe Text

Replication Configuration template EBS encryption key ARN.

defaultLargeStagingDiskType :: Maybe ReplicationConfigurationDefaultLargeStagingDiskType

Replication Configuration template use dedault large Staging Disk type.

bandwidthThrottling :: Maybe Natural

Replication Configuration template bandwidth throtting.

dataPlaneRouting :: Maybe ReplicationConfigurationDataPlaneRouting

Replication Configuration template data plane routing.

useDedicatedReplicationServer :: Maybe Bool

Replication Configuration template use Dedicated Replication Server.

tags :: Maybe (Sensitive (HashMap Text Text))

Replication Configuration template Tags.

replicationConfigurationTemplateID :: Text

Replication Configuration template template ID.

newReplicationConfigurationTemplate

Arguments

:: Text
$sel:replicationConfigurationTemplateID:ReplicationConfigurationTemplate'

-> ReplicationConfigurationTemplate

Create a value of ReplicationConfigurationTemplate with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:createPublicIP:ReplicationConfigurationTemplate', replicationConfigurationTemplate_createPublicIP - Replication Configuration template create Public IP.

$sel:stagingAreaTags:ReplicationConfigurationTemplate', replicationConfigurationTemplate_stagingAreaTags - Replication Configuration template Staging Area Tags.

$sel:arn:ReplicationConfigurationTemplate', replicationConfigurationTemplate_arn - Replication Configuration template ARN.

$sel:stagingAreaSubnetId:ReplicationConfigurationTemplate', replicationConfigurationTemplate_stagingAreaSubnetId - Replication Configuration template Staging Area subnet ID.

$sel:replicationServerInstanceType:ReplicationConfigurationTemplate', replicationConfigurationTemplate_replicationServerInstanceType - Replication Configuration template server instance type.

$sel:ebsEncryption:ReplicationConfigurationTemplate', replicationConfigurationTemplate_ebsEncryption - Replication Configuration template EBS encryption.

$sel:associateDefaultSecurityGroup:ReplicationConfigurationTemplate', replicationConfigurationTemplate_associateDefaultSecurityGroup - Replication Configuration template associate default Application Migration Service Security group.

$sel:replicationServersSecurityGroupsIDs:ReplicationConfigurationTemplate', replicationConfigurationTemplate_replicationServersSecurityGroupsIDs - Replication Configuration template server Security Groups IDs.

$sel:ebsEncryptionKeyArn:ReplicationConfigurationTemplate', replicationConfigurationTemplate_ebsEncryptionKeyArn - Replication Configuration template EBS encryption key ARN.

$sel:defaultLargeStagingDiskType:ReplicationConfigurationTemplate', replicationConfigurationTemplate_defaultLargeStagingDiskType - Replication Configuration template use dedault large Staging Disk type.

$sel:bandwidthThrottling:ReplicationConfigurationTemplate', replicationConfigurationTemplate_bandwidthThrottling - Replication Configuration template bandwidth throtting.

$sel:dataPlaneRouting:ReplicationConfigurationTemplate', replicationConfigurationTemplate_dataPlaneRouting - Replication Configuration template data plane routing.

$sel:useDedicatedReplicationServer:ReplicationConfigurationTemplate', replicationConfigurationTemplate_useDedicatedReplicationServer - Replication Configuration template use Dedicated Replication Server.

$sel:tags:ReplicationConfigurationTemplate', replicationConfigurationTemplate_tags - Replication Configuration template Tags.

$sel:replicationConfigurationTemplateID:ReplicationConfigurationTemplate', replicationConfigurationTemplate_replicationConfigurationTemplateID - Replication Configuration template template ID.

Response Lenses

replicationConfigurationTemplate_createPublicIP :: Lens' ReplicationConfigurationTemplate (Maybe Bool)

Replication Configuration template create Public IP.

replicationConfigurationTemplate_stagingAreaTags :: Lens' ReplicationConfigurationTemplate (Maybe (HashMap Text Text))

Replication Configuration template Staging Area Tags.

replicationConfigurationTemplate_arn :: Lens' ReplicationConfigurationTemplate (Maybe Text)

Replication Configuration template ARN.

replicationConfigurationTemplate_stagingAreaSubnetId :: Lens' ReplicationConfigurationTemplate (Maybe Text)

Replication Configuration template Staging Area subnet ID.

replicationConfigurationTemplate_replicationServerInstanceType :: Lens' ReplicationConfigurationTemplate (Maybe Text)

Replication Configuration template server instance type.

replicationConfigurationTemplate_ebsEncryption :: Lens' ReplicationConfigurationTemplate (Maybe ReplicationConfigurationEbsEncryption)

Replication Configuration template EBS encryption.

replicationConfigurationTemplate_associateDefaultSecurityGroup :: Lens' ReplicationConfigurationTemplate (Maybe Bool)

Replication Configuration template associate default Application Migration Service Security group.

replicationConfigurationTemplate_replicationServersSecurityGroupsIDs :: Lens' ReplicationConfigurationTemplate (Maybe [Text])

Replication Configuration template server Security Groups IDs.

replicationConfigurationTemplate_ebsEncryptionKeyArn :: Lens' ReplicationConfigurationTemplate (Maybe Text)

Replication Configuration template EBS encryption key ARN.

replicationConfigurationTemplate_defaultLargeStagingDiskType :: Lens' ReplicationConfigurationTemplate (Maybe ReplicationConfigurationDefaultLargeStagingDiskType)

Replication Configuration template use dedault large Staging Disk type.

replicationConfigurationTemplate_bandwidthThrottling :: Lens' ReplicationConfigurationTemplate (Maybe Natural)

Replication Configuration template bandwidth throtting.

replicationConfigurationTemplate_dataPlaneRouting :: Lens' ReplicationConfigurationTemplate (Maybe ReplicationConfigurationDataPlaneRouting)

Replication Configuration template data plane routing.

replicationConfigurationTemplate_useDedicatedReplicationServer :: Lens' ReplicationConfigurationTemplate (Maybe Bool)

Replication Configuration template use Dedicated Replication Server.

replicationConfigurationTemplate_tags :: Lens' ReplicationConfigurationTemplate (Maybe (HashMap Text Text))

Replication Configuration template Tags.

replicationConfigurationTemplate_replicationConfigurationTemplateID :: Lens' ReplicationConfigurationTemplate Text

Replication Configuration template template ID.