Amazonka.MQ.CreateBroker

Description

Creates a broker. Note: This API is asynchronous.

To create a broker, you must either use the AmazonMQFullAccess IAM policy or include the following EC2 permissions in your IAM policy.

ec2:CreateNetworkInterface

This permission is required to allow Amazon MQ to create an elastic network interface (ENI) on behalf of your account.

ec2:CreateNetworkInterfacePermission

This permission is required to attach the ENI to the broker instance.

ec2:DeleteNetworkInterface

ec2:DeleteNetworkInterfacePermission

ec2:DetachNetworkInterface

ec2:DescribeInternetGateways

ec2:DescribeNetworkInterfaces

ec2:DescribeNetworkInterfacePermissions

ec2:DescribeRouteTables

ec2:DescribeSecurityGroups

ec2:DescribeSubnets

ec2:DescribeVpcs

For more information, see Create an IAM User and Get Your AWS Credentials and Never Modify or Delete the Amazon MQ Elastic Network Interface in the Amazon MQ Developer Guide.

Creating a Request

data CreateBroker

Creates a broker using the specified properties.

See: newCreateBroker smart constructor.

Constructors

CreateBroker'

Fields

securityGroups :: Maybe [Text]

The list of rules (1 minimum, 125 maximum) that authorize connections to brokers.

subnetIds :: Maybe [Text]

The list of groups that define which subnets and IP ranges the broker can use from different Availability Zones. If you specify more than one subnet, the subnets must be in different Availability Zones. Amazon MQ will not be able to create VPC endpoints for your broker with multiple subnets in the same Availability Zone. A SINGLE_INSTANCE deployment requires one subnet (for example, the default subnet). An ACTIVE_STANDBY_MULTI_AZ Amazon MQ for ActiveMQ deployment requires two subnets. A CLUSTER_MULTI_AZ Amazon MQ for RabbitMQ deployment has no subnet requirements when deployed with public accessibility. Deployment without public accessibility requires at least one subnet.

If you specify subnets in a shared VPC for a RabbitMQ broker, the associated VPC to which the specified subnets belong must be owned by your AWS account. Amazon MQ will not be able to create VPC endpoints in VPCs that are not owned by your AWS account.

creatorRequestId :: Maybe Text

The unique ID that the requester receives for the created broker. Amazon MQ passes your ID with the API action. Note: We recommend using a Universally Unique Identifier (UUID) for the creatorRequestId. You may omit the creatorRequestId if your application doesn't require idempotency.

authenticationStrategy :: Maybe AuthenticationStrategy

Optional. The authentication strategy used to secure the broker. The default is SIMPLE.

ldapServerMetadata :: Maybe LdapServerMetadataInput

Optional. The metadata of the LDAP server used to authenticate and authorize connections to the broker. Does not apply to RabbitMQ brokers.

maintenanceWindowStartTime :: Maybe WeeklyStartTime

The parameters that determine the WeeklyStartTime.

logs :: Maybe Logs

Enables Amazon CloudWatch logging for brokers.

encryptionOptions :: Maybe EncryptionOptions

Encryption options for the broker. Does not apply to RabbitMQ brokers.

configuration :: Maybe ConfigurationId

A list of information about the configuration.

tags :: Maybe (HashMap Text Text)

Create tags when creating the broker.

storageType :: Maybe BrokerStorageType

The broker's storage type.

engineVersion :: Text

Required. The broker engine's version. For a list of supported engine versions, see Supported engines.

hostInstanceType :: Text

Required. The broker's instance type.

autoMinorVersionUpgrade :: Bool

Enables automatic upgrades to new minor versions for brokers, as new versions are released and supported by Amazon MQ. Automatic upgrades occur during the scheduled maintenance window of the broker or after a manual broker reboot. Set to true by default, if no value is specified.

users :: [User]

Required. The list of broker users (persons or applications) who can access queues and topics. This value can contain only alphanumeric characters, dashes, periods, underscores, and tildes (- . _ ~). This value must be 2-100 characters long.

Amazon MQ for RabbitMQ

When you create an Amazon MQ for RabbitMQ broker, one and only one administrative user is accepted and created when a broker is first provisioned. All subsequent broker users are created by making RabbitMQ API calls directly to brokers or via the RabbitMQ web console.

brokerName :: Text

Required. The broker's name. This value must be unique in your AWS account, 1-50 characters long, must contain only letters, numbers, dashes, and underscores, and must not contain white spaces, brackets, wildcard characters, or special characters.

deploymentMode :: DeploymentMode

Required. The broker's deployment mode.

engineType :: EngineType

Required. The type of broker engine. Currently, Amazon MQ supports ACTIVEMQ and RABBITMQ.

publiclyAccessible :: Bool

Enables connections from applications outside of the VPC that hosts the broker's subnets. Set to false by default, if no value is provided.

newCreateBroker

Arguments

:: Text
$sel:engineVersion:CreateBroker'

-> Text
$sel:hostInstanceType:CreateBroker'

-> Bool
$sel:autoMinorVersionUpgrade:CreateBroker'

-> Text
$sel:brokerName:CreateBroker'

-> DeploymentMode
$sel:deploymentMode:CreateBroker'

-> EngineType
$sel:engineType:CreateBroker'

-> Bool
$sel:publiclyAccessible:CreateBroker'

-> CreateBroker

Create a value of CreateBroker with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:securityGroups:CreateBroker', createBroker_securityGroups - The list of rules (1 minimum, 125 maximum) that authorize connections to brokers.

$sel:subnetIds:CreateBroker', createBroker_subnetIds - The list of groups that define which subnets and IP ranges the broker can use from different Availability Zones. If you specify more than one subnet, the subnets must be in different Availability Zones. Amazon MQ will not be able to create VPC endpoints for your broker with multiple subnets in the same Availability Zone. A SINGLE_INSTANCE deployment requires one subnet (for example, the default subnet). An ACTIVE_STANDBY_MULTI_AZ Amazon MQ for ActiveMQ deployment requires two subnets. A CLUSTER_MULTI_AZ Amazon MQ for RabbitMQ deployment has no subnet requirements when deployed with public accessibility. Deployment without public accessibility requires at least one subnet.

If you specify subnets in a shared VPC for a RabbitMQ broker, the associated VPC to which the specified subnets belong must be owned by your AWS account. Amazon MQ will not be able to create VPC endpoints in VPCs that are not owned by your AWS account.

$sel:creatorRequestId:CreateBroker', createBroker_creatorRequestId - The unique ID that the requester receives for the created broker. Amazon MQ passes your ID with the API action. Note: We recommend using a Universally Unique Identifier (UUID) for the creatorRequestId. You may omit the creatorRequestId if your application doesn't require idempotency.

$sel:authenticationStrategy:CreateBroker', createBroker_authenticationStrategy - Optional. The authentication strategy used to secure the broker. The default is SIMPLE.

$sel:ldapServerMetadata:CreateBroker', createBroker_ldapServerMetadata - Optional. The metadata of the LDAP server used to authenticate and authorize connections to the broker. Does not apply to RabbitMQ brokers.

$sel:maintenanceWindowStartTime:CreateBroker', createBroker_maintenanceWindowStartTime - The parameters that determine the WeeklyStartTime.

$sel:logs:CreateBroker', createBroker_logs - Enables Amazon CloudWatch logging for brokers.

$sel:encryptionOptions:CreateBroker', createBroker_encryptionOptions - Encryption options for the broker. Does not apply to RabbitMQ brokers.

$sel:configuration:CreateBroker', createBroker_configuration - A list of information about the configuration.

$sel:tags:CreateBroker', createBroker_tags - Create tags when creating the broker.

$sel:storageType:CreateBroker', createBroker_storageType - The broker's storage type.

$sel:engineVersion:CreateBroker', createBroker_engineVersion - Required. The broker engine's version. For a list of supported engine versions, see Supported engines.

$sel:hostInstanceType:CreateBroker', createBroker_hostInstanceType - Required. The broker's instance type.

$sel:autoMinorVersionUpgrade:CreateBroker', createBroker_autoMinorVersionUpgrade - Enables automatic upgrades to new minor versions for brokers, as new versions are released and supported by Amazon MQ. Automatic upgrades occur during the scheduled maintenance window of the broker or after a manual broker reboot. Set to true by default, if no value is specified.

$sel:users:CreateBroker', createBroker_users - Required. The list of broker users (persons or applications) who can access queues and topics. This value can contain only alphanumeric characters, dashes, periods, underscores, and tildes (- . _ ~). This value must be 2-100 characters long.

Amazon MQ for RabbitMQ

When you create an Amazon MQ for RabbitMQ broker, one and only one administrative user is accepted and created when a broker is first provisioned. All subsequent broker users are created by making RabbitMQ API calls directly to brokers or via the RabbitMQ web console.

$sel:brokerName:CreateBroker', createBroker_brokerName - Required. The broker's name. This value must be unique in your AWS account, 1-50 characters long, must contain only letters, numbers, dashes, and underscores, and must not contain white spaces, brackets, wildcard characters, or special characters.

$sel:deploymentMode:CreateBroker', createBroker_deploymentMode - Required. The broker's deployment mode.

$sel:engineType:CreateBroker', createBroker_engineType - Required. The type of broker engine. Currently, Amazon MQ supports ACTIVEMQ and RABBITMQ.

$sel:publiclyAccessible:CreateBroker', createBroker_publiclyAccessible - Enables connections from applications outside of the VPC that hosts the broker's subnets. Set to false by default, if no value is provided.

Request Lenses

createBroker_securityGroups :: Lens' CreateBroker (Maybe [Text])

The list of rules (1 minimum, 125 maximum) that authorize connections to brokers.

createBroker_subnetIds :: Lens' CreateBroker (Maybe [Text])

The list of groups that define which subnets and IP ranges the broker can use from different Availability Zones. If you specify more than one subnet, the subnets must be in different Availability Zones. Amazon MQ will not be able to create VPC endpoints for your broker with multiple subnets in the same Availability Zone. A SINGLE_INSTANCE deployment requires one subnet (for example, the default subnet). An ACTIVE_STANDBY_MULTI_AZ Amazon MQ for ActiveMQ deployment requires two subnets. A CLUSTER_MULTI_AZ Amazon MQ for RabbitMQ deployment has no subnet requirements when deployed with public accessibility. Deployment without public accessibility requires at least one subnet.

If you specify subnets in a shared VPC for a RabbitMQ broker, the associated VPC to which the specified subnets belong must be owned by your AWS account. Amazon MQ will not be able to create VPC endpoints in VPCs that are not owned by your AWS account.

createBroker_creatorRequestId :: Lens' CreateBroker (Maybe Text)

The unique ID that the requester receives for the created broker. Amazon MQ passes your ID with the API action. Note: We recommend using a Universally Unique Identifier (UUID) for the creatorRequestId. You may omit the creatorRequestId if your application doesn't require idempotency.

createBroker_authenticationStrategy :: Lens' CreateBroker (Maybe AuthenticationStrategy)

Optional. The authentication strategy used to secure the broker. The default is SIMPLE.

createBroker_ldapServerMetadata :: Lens' CreateBroker (Maybe LdapServerMetadataInput)

Optional. The metadata of the LDAP server used to authenticate and authorize connections to the broker. Does not apply to RabbitMQ brokers.

createBroker_maintenanceWindowStartTime :: Lens' CreateBroker (Maybe WeeklyStartTime)

The parameters that determine the WeeklyStartTime.

createBroker_logs :: Lens' CreateBroker (Maybe Logs)

Enables Amazon CloudWatch logging for brokers.

createBroker_encryptionOptions :: Lens' CreateBroker (Maybe EncryptionOptions)

Encryption options for the broker. Does not apply to RabbitMQ brokers.

createBroker_configuration :: Lens' CreateBroker (Maybe ConfigurationId)

A list of information about the configuration.

createBroker_tags :: Lens' CreateBroker (Maybe (HashMap Text Text))

Create tags when creating the broker.

createBroker_storageType :: Lens' CreateBroker (Maybe BrokerStorageType)

The broker's storage type.

createBroker_engineVersion :: Lens' CreateBroker Text

Required. The broker engine's version. For a list of supported engine versions, see Supported engines.

createBroker_hostInstanceType :: Lens' CreateBroker Text

Required. The broker's instance type.

createBroker_autoMinorVersionUpgrade :: Lens' CreateBroker Bool

Enables automatic upgrades to new minor versions for brokers, as new versions are released and supported by Amazon MQ. Automatic upgrades occur during the scheduled maintenance window of the broker or after a manual broker reboot. Set to true by default, if no value is specified.

createBroker_users :: Lens' CreateBroker [User]

Required. The list of broker users (persons or applications) who can access queues and topics. This value can contain only alphanumeric characters, dashes, periods, underscores, and tildes (- . _ ~). This value must be 2-100 characters long.

Amazon MQ for RabbitMQ

When you create an Amazon MQ for RabbitMQ broker, one and only one administrative user is accepted and created when a broker is first provisioned. All subsequent broker users are created by making RabbitMQ API calls directly to brokers or via the RabbitMQ web console.

createBroker_brokerName :: Lens' CreateBroker Text

Required. The broker's name. This value must be unique in your AWS account, 1-50 characters long, must contain only letters, numbers, dashes, and underscores, and must not contain white spaces, brackets, wildcard characters, or special characters.

createBroker_deploymentMode :: Lens' CreateBroker DeploymentMode

Required. The broker's deployment mode.

createBroker_engineType :: Lens' CreateBroker EngineType

Required. The type of broker engine. Currently, Amazon MQ supports ACTIVEMQ and RABBITMQ.

createBroker_publiclyAccessible :: Lens' CreateBroker Bool

Enables connections from applications outside of the VPC that hosts the broker's subnets. Set to false by default, if no value is provided.

Destructuring the Response

data CreateBrokerResponse

See: newCreateBrokerResponse smart constructor.

Constructors

CreateBrokerResponse'

Fields

brokerId :: Maybe Text

The unique ID that Amazon MQ generates for the broker.

brokerArn :: Maybe Text

The broker's Amazon Resource Name (ARN).

httpStatus :: Int

The response's http status code.

newCreateBrokerResponse

Arguments

:: Int
$sel:httpStatus:CreateBrokerResponse'

-> CreateBrokerResponse

Create a value of CreateBrokerResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:brokerId:CreateBrokerResponse', createBrokerResponse_brokerId - The unique ID that Amazon MQ generates for the broker.

$sel:brokerArn:CreateBrokerResponse', createBrokerResponse_brokerArn - The broker's Amazon Resource Name (ARN).

$sel:httpStatus:CreateBrokerResponse', createBrokerResponse_httpStatus - The response's http status code.

Response Lenses

createBrokerResponse_brokerId :: Lens' CreateBrokerResponse (Maybe Text)

The unique ID that Amazon MQ generates for the broker.

createBrokerResponse_brokerArn :: Lens' CreateBrokerResponse (Maybe Text)

The broker's Amazon Resource Name (ARN).

createBrokerResponse_httpStatus :: Lens' CreateBrokerResponse Int

The response's http status code.