Amazonka.MQ.DescribeBroker

Description

Returns information about the specified broker.

Creating a Request

data DescribeBroker

See: newDescribeBroker smart constructor.

Constructors

DescribeBroker'

Fields

brokerId :: Text

The unique ID that Amazon MQ generates for the broker.

newDescribeBroker

Arguments

:: Text
$sel:brokerId:DescribeBroker'

-> DescribeBroker

Create a value of DescribeBroker with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:brokerId:DescribeBroker', describeBroker_brokerId - The unique ID that Amazon MQ generates for the broker.

Request Lenses

describeBroker_brokerId :: Lens' DescribeBroker Text

The unique ID that Amazon MQ generates for the broker.

Destructuring the Response

data DescribeBrokerResponse

See: newDescribeBrokerResponse smart constructor.

Constructors

DescribeBrokerResponse'

Fields

brokerName :: Maybe Text

The broker's name. This value must be unique in your AWS account, 1-50 characters long, must contain only letters, numbers, dashes, and underscores, and must not contain white spaces, brackets, wildcard characters, or special characters.

engineVersion :: Maybe Text

The broker engine's version. For a list of supported engine versions, see Supported engines.

pendingAuthenticationStrategy :: Maybe AuthenticationStrategy

The authentication strategy that will be applied when the broker is rebooted. The default is SIMPLE.

brokerState :: Maybe BrokerState

The broker's status.

publiclyAccessible :: Maybe Bool

Enables connections from applications outside of the VPC that hosts the broker's subnets.

autoMinorVersionUpgrade :: Maybe Bool

Enables automatic upgrades to new minor versions for brokers, as new versions are released and supported by Amazon MQ. Automatic upgrades occur during the scheduled maintenance window of the broker or after a manual broker reboot.

securityGroups :: Maybe [Text]

The list of rules (1 minimum, 125 maximum) that authorize connections to brokers.

users :: Maybe [UserSummary]

The list of all broker usernames for the specified broker.

pendingSecurityGroups :: Maybe [Text]

The list of pending security groups to authorize connections to brokers.

subnetIds :: Maybe [Text]

The list of groups that define which subnets and IP ranges the broker can use from different Availability Zones.

created :: Maybe POSIX

The time when the broker was created.

configurations :: Maybe Configurations

The list of all revisions for the specified configuration.

authenticationStrategy :: Maybe AuthenticationStrategy

The authentication strategy used to secure the broker. The default is SIMPLE.

pendingHostInstanceType :: Maybe Text

The broker's host instance type to upgrade to. For a list of supported instance types, see Broker instance types.

ldapServerMetadata :: Maybe LdapServerMetadataOutput

The metadata of the LDAP server used to authenticate and authorize connections to the broker.

maintenanceWindowStartTime :: Maybe WeeklyStartTime

The parameters that determine the WeeklyStartTime.

logs :: Maybe LogsSummary

The list of information about logs currently enabled and pending to be deployed for the specified broker.

encryptionOptions :: Maybe EncryptionOptions

Encryption options for the broker. Does not apply to RabbitMQ brokers.

deploymentMode :: Maybe DeploymentMode

The broker's deployment mode.

pendingEngineVersion :: Maybe Text

The broker engine version to upgrade to. For a list of supported engine versions, see Supported engines.

brokerId :: Maybe Text

The unique ID that Amazon MQ generates for the broker.

pendingLdapServerMetadata :: Maybe LdapServerMetadataOutput

The metadata of the LDAP server that will be used to authenticate and authorize connections to the broker after it is rebooted.

engineType :: Maybe EngineType

The type of broker engine. Currently, Amazon MQ supports ACTIVEMQ and RABBITMQ.

brokerArn :: Maybe Text

The broker's Amazon Resource Name (ARN).

tags :: Maybe (HashMap Text Text)

The list of all tags associated with this broker.

brokerInstances :: Maybe [BrokerInstance]

A list of information about allocated brokers.

hostInstanceType :: Maybe Text

The broker's instance type.

storageType :: Maybe BrokerStorageType

The broker's storage type.

httpStatus :: Int

The response's http status code.

newDescribeBrokerResponse

Arguments

:: Int
$sel:httpStatus:DescribeBrokerResponse'

-> DescribeBrokerResponse

Create a value of DescribeBrokerResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:brokerName:DescribeBrokerResponse', describeBrokerResponse_brokerName - The broker's name. This value must be unique in your AWS account, 1-50 characters long, must contain only letters, numbers, dashes, and underscores, and must not contain white spaces, brackets, wildcard characters, or special characters.

$sel:engineVersion:DescribeBrokerResponse', describeBrokerResponse_engineVersion - The broker engine's version. For a list of supported engine versions, see Supported engines.

$sel:pendingAuthenticationStrategy:DescribeBrokerResponse', describeBrokerResponse_pendingAuthenticationStrategy - The authentication strategy that will be applied when the broker is rebooted. The default is SIMPLE.

$sel:brokerState:DescribeBrokerResponse', describeBrokerResponse_brokerState - The broker's status.

$sel:publiclyAccessible:DescribeBrokerResponse', describeBrokerResponse_publiclyAccessible - Enables connections from applications outside of the VPC that hosts the broker's subnets.

$sel:autoMinorVersionUpgrade:DescribeBrokerResponse', describeBrokerResponse_autoMinorVersionUpgrade - Enables automatic upgrades to new minor versions for brokers, as new versions are released and supported by Amazon MQ. Automatic upgrades occur during the scheduled maintenance window of the broker or after a manual broker reboot.

$sel:securityGroups:DescribeBrokerResponse', describeBrokerResponse_securityGroups - The list of rules (1 minimum, 125 maximum) that authorize connections to brokers.

$sel:users:DescribeBrokerResponse', describeBrokerResponse_users - The list of all broker usernames for the specified broker.

$sel:pendingSecurityGroups:DescribeBrokerResponse', describeBrokerResponse_pendingSecurityGroups - The list of pending security groups to authorize connections to brokers.

$sel:subnetIds:DescribeBrokerResponse', describeBrokerResponse_subnetIds - The list of groups that define which subnets and IP ranges the broker can use from different Availability Zones.

$sel:created:DescribeBrokerResponse', describeBrokerResponse_created - The time when the broker was created.

$sel:configurations:DescribeBrokerResponse', describeBrokerResponse_configurations - The list of all revisions for the specified configuration.

$sel:authenticationStrategy:DescribeBrokerResponse', describeBrokerResponse_authenticationStrategy - The authentication strategy used to secure the broker. The default is SIMPLE.

$sel:pendingHostInstanceType:DescribeBrokerResponse', describeBrokerResponse_pendingHostInstanceType - The broker's host instance type to upgrade to. For a list of supported instance types, see Broker instance types.

$sel:ldapServerMetadata:DescribeBrokerResponse', describeBrokerResponse_ldapServerMetadata - The metadata of the LDAP server used to authenticate and authorize connections to the broker.

$sel:maintenanceWindowStartTime:DescribeBrokerResponse', describeBrokerResponse_maintenanceWindowStartTime - The parameters that determine the WeeklyStartTime.

$sel:logs:DescribeBrokerResponse', describeBrokerResponse_logs - The list of information about logs currently enabled and pending to be deployed for the specified broker.

$sel:encryptionOptions:DescribeBrokerResponse', describeBrokerResponse_encryptionOptions - Encryption options for the broker. Does not apply to RabbitMQ brokers.

$sel:deploymentMode:DescribeBrokerResponse', describeBrokerResponse_deploymentMode - The broker's deployment mode.

$sel:pendingEngineVersion:DescribeBrokerResponse', describeBrokerResponse_pendingEngineVersion - The broker engine version to upgrade to. For a list of supported engine versions, see Supported engines.

$sel:brokerId:DescribeBroker', describeBrokerResponse_brokerId - The unique ID that Amazon MQ generates for the broker.

$sel:pendingLdapServerMetadata:DescribeBrokerResponse', describeBrokerResponse_pendingLdapServerMetadata - The metadata of the LDAP server that will be used to authenticate and authorize connections to the broker after it is rebooted.

$sel:engineType:DescribeBrokerResponse', describeBrokerResponse_engineType - The type of broker engine. Currently, Amazon MQ supports ACTIVEMQ and RABBITMQ.

$sel:brokerArn:DescribeBrokerResponse', describeBrokerResponse_brokerArn - The broker's Amazon Resource Name (ARN).

$sel:tags:DescribeBrokerResponse', describeBrokerResponse_tags - The list of all tags associated with this broker.

$sel:brokerInstances:DescribeBrokerResponse', describeBrokerResponse_brokerInstances - A list of information about allocated brokers.

$sel:hostInstanceType:DescribeBrokerResponse', describeBrokerResponse_hostInstanceType - The broker's instance type.

$sel:storageType:DescribeBrokerResponse', describeBrokerResponse_storageType - The broker's storage type.

$sel:httpStatus:DescribeBrokerResponse', describeBrokerResponse_httpStatus - The response's http status code.

Response Lenses

describeBrokerResponse_brokerName :: Lens' DescribeBrokerResponse (Maybe Text)

The broker's name. This value must be unique in your AWS account, 1-50 characters long, must contain only letters, numbers, dashes, and underscores, and must not contain white spaces, brackets, wildcard characters, or special characters.

describeBrokerResponse_engineVersion :: Lens' DescribeBrokerResponse (Maybe Text)

The broker engine's version. For a list of supported engine versions, see Supported engines.

describeBrokerResponse_pendingAuthenticationStrategy :: Lens' DescribeBrokerResponse (Maybe AuthenticationStrategy)

The authentication strategy that will be applied when the broker is rebooted. The default is SIMPLE.

describeBrokerResponse_brokerState :: Lens' DescribeBrokerResponse (Maybe BrokerState)

The broker's status.

describeBrokerResponse_publiclyAccessible :: Lens' DescribeBrokerResponse (Maybe Bool)

Enables connections from applications outside of the VPC that hosts the broker's subnets.

describeBrokerResponse_autoMinorVersionUpgrade :: Lens' DescribeBrokerResponse (Maybe Bool)

Enables automatic upgrades to new minor versions for brokers, as new versions are released and supported by Amazon MQ. Automatic upgrades occur during the scheduled maintenance window of the broker or after a manual broker reboot.

describeBrokerResponse_securityGroups :: Lens' DescribeBrokerResponse (Maybe [Text])

The list of rules (1 minimum, 125 maximum) that authorize connections to brokers.

describeBrokerResponse_users :: Lens' DescribeBrokerResponse (Maybe [UserSummary])

The list of all broker usernames for the specified broker.

describeBrokerResponse_pendingSecurityGroups :: Lens' DescribeBrokerResponse (Maybe [Text])

The list of pending security groups to authorize connections to brokers.

describeBrokerResponse_subnetIds :: Lens' DescribeBrokerResponse (Maybe [Text])

The list of groups that define which subnets and IP ranges the broker can use from different Availability Zones.

describeBrokerResponse_created :: Lens' DescribeBrokerResponse (Maybe UTCTime)

The time when the broker was created.

describeBrokerResponse_configurations :: Lens' DescribeBrokerResponse (Maybe Configurations)

The list of all revisions for the specified configuration.

describeBrokerResponse_authenticationStrategy :: Lens' DescribeBrokerResponse (Maybe AuthenticationStrategy)

The authentication strategy used to secure the broker. The default is SIMPLE.

describeBrokerResponse_pendingHostInstanceType :: Lens' DescribeBrokerResponse (Maybe Text)

The broker's host instance type to upgrade to. For a list of supported instance types, see Broker instance types.

describeBrokerResponse_ldapServerMetadata :: Lens' DescribeBrokerResponse (Maybe LdapServerMetadataOutput)

The metadata of the LDAP server used to authenticate and authorize connections to the broker.

describeBrokerResponse_maintenanceWindowStartTime :: Lens' DescribeBrokerResponse (Maybe WeeklyStartTime)

The parameters that determine the WeeklyStartTime.

describeBrokerResponse_logs :: Lens' DescribeBrokerResponse (Maybe LogsSummary)

The list of information about logs currently enabled and pending to be deployed for the specified broker.

describeBrokerResponse_encryptionOptions :: Lens' DescribeBrokerResponse (Maybe EncryptionOptions)

Encryption options for the broker. Does not apply to RabbitMQ brokers.

describeBrokerResponse_deploymentMode :: Lens' DescribeBrokerResponse (Maybe DeploymentMode)

The broker's deployment mode.

describeBrokerResponse_pendingEngineVersion :: Lens' DescribeBrokerResponse (Maybe Text)

The broker engine version to upgrade to. For a list of supported engine versions, see Supported engines.

describeBrokerResponse_brokerId :: Lens' DescribeBrokerResponse (Maybe Text)

The unique ID that Amazon MQ generates for the broker.

describeBrokerResponse_pendingLdapServerMetadata :: Lens' DescribeBrokerResponse (Maybe LdapServerMetadataOutput)

The metadata of the LDAP server that will be used to authenticate and authorize connections to the broker after it is rebooted.

describeBrokerResponse_engineType :: Lens' DescribeBrokerResponse (Maybe EngineType)

The type of broker engine. Currently, Amazon MQ supports ACTIVEMQ and RABBITMQ.

describeBrokerResponse_brokerArn :: Lens' DescribeBrokerResponse (Maybe Text)

The broker's Amazon Resource Name (ARN).

describeBrokerResponse_tags :: Lens' DescribeBrokerResponse (Maybe (HashMap Text Text))

The list of all tags associated with this broker.

describeBrokerResponse_brokerInstances :: Lens' DescribeBrokerResponse (Maybe [BrokerInstance])

A list of information about allocated brokers.

describeBrokerResponse_hostInstanceType :: Lens' DescribeBrokerResponse (Maybe Text)

The broker's instance type.

describeBrokerResponse_storageType :: Lens' DescribeBrokerResponse (Maybe BrokerStorageType)

The broker's storage type.

describeBrokerResponse_httpStatus :: Lens' DescribeBrokerResponse Int

The response's http status code.