Amazonka.MQ.DescribeConfiguration

Description

Returns information about the specified configuration.

Creating a Request

data DescribeConfiguration

See: newDescribeConfiguration smart constructor.

Constructors

DescribeConfiguration'

Fields

configurationId :: Text

The unique ID that Amazon MQ generates for the configuration.

newDescribeConfiguration

Arguments

:: Text
$sel:configurationId:DescribeConfiguration'

-> DescribeConfiguration

Create a value of DescribeConfiguration with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:configurationId:DescribeConfiguration', describeConfiguration_configurationId - The unique ID that Amazon MQ generates for the configuration.

Request Lenses

describeConfiguration_configurationId :: Lens' DescribeConfiguration Text

The unique ID that Amazon MQ generates for the configuration.

Destructuring the Response

data DescribeConfigurationResponse

See: newDescribeConfigurationResponse smart constructor.

Constructors

DescribeConfigurationResponse'

Fields

engineVersion :: Maybe Text

Required. The broker engine's version. For a list of supported engine versions, see, Supported engines.

arn :: Maybe Text

Required. The ARN of the configuration.

latestRevision :: Maybe ConfigurationRevision

Required. The latest revision of the configuration.

created :: Maybe POSIX

Required. The date and time of the configuration revision.

authenticationStrategy :: Maybe AuthenticationStrategy

Optional. The authentication strategy associated with the configuration. The default is SIMPLE.

name :: Maybe Text

Required. The name of the configuration. This value can contain only alphanumeric characters, dashes, periods, underscores, and tildes (- . _ ~). This value must be 1-150 characters long.

id :: Maybe Text

Required. The unique ID that Amazon MQ generates for the configuration.

description :: Maybe Text

Required. The description of the configuration.

engineType :: Maybe EngineType

Required. The type of broker engine. Currently, Amazon MQ supports ACTIVEMQ and RABBITMQ.

tags :: Maybe (HashMap Text Text)

The list of all tags associated with this configuration.

httpStatus :: Int

The response's http status code.

newDescribeConfigurationResponse

Arguments

:: Int
$sel:httpStatus:DescribeConfigurationResponse'

-> DescribeConfigurationResponse

Create a value of DescribeConfigurationResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:engineVersion:DescribeConfigurationResponse', describeConfigurationResponse_engineVersion - Required. The broker engine's version. For a list of supported engine versions, see, Supported engines.

$sel:arn:DescribeConfigurationResponse', describeConfigurationResponse_arn - Required. The ARN of the configuration.

$sel:latestRevision:DescribeConfigurationResponse', describeConfigurationResponse_latestRevision - Required. The latest revision of the configuration.

$sel:created:DescribeConfigurationResponse', describeConfigurationResponse_created - Required. The date and time of the configuration revision.

$sel:authenticationStrategy:DescribeConfigurationResponse', describeConfigurationResponse_authenticationStrategy - Optional. The authentication strategy associated with the configuration. The default is SIMPLE.

$sel:name:DescribeConfigurationResponse', describeConfigurationResponse_name - Required. The name of the configuration. This value can contain only alphanumeric characters, dashes, periods, underscores, and tildes (- . _ ~). This value must be 1-150 characters long.

$sel:id:DescribeConfigurationResponse', describeConfigurationResponse_id - Required. The unique ID that Amazon MQ generates for the configuration.

$sel:description:DescribeConfigurationResponse', describeConfigurationResponse_description - Required. The description of the configuration.

$sel:engineType:DescribeConfigurationResponse', describeConfigurationResponse_engineType - Required. The type of broker engine. Currently, Amazon MQ supports ACTIVEMQ and RABBITMQ.

$sel:tags:DescribeConfigurationResponse', describeConfigurationResponse_tags - The list of all tags associated with this configuration.

$sel:httpStatus:DescribeConfigurationResponse', describeConfigurationResponse_httpStatus - The response's http status code.

Response Lenses

describeConfigurationResponse_engineVersion :: Lens' DescribeConfigurationResponse (Maybe Text)

Required. The broker engine's version. For a list of supported engine versions, see, Supported engines.

describeConfigurationResponse_arn :: Lens' DescribeConfigurationResponse (Maybe Text)

Required. The ARN of the configuration.

describeConfigurationResponse_latestRevision :: Lens' DescribeConfigurationResponse (Maybe ConfigurationRevision)

Required. The latest revision of the configuration.

describeConfigurationResponse_created :: Lens' DescribeConfigurationResponse (Maybe UTCTime)

Required. The date and time of the configuration revision.

describeConfigurationResponse_authenticationStrategy :: Lens' DescribeConfigurationResponse (Maybe AuthenticationStrategy)

Optional. The authentication strategy associated with the configuration. The default is SIMPLE.

describeConfigurationResponse_name :: Lens' DescribeConfigurationResponse (Maybe Text)

Required. The name of the configuration. This value can contain only alphanumeric characters, dashes, periods, underscores, and tildes (- . _ ~). This value must be 1-150 characters long.

describeConfigurationResponse_id :: Lens' DescribeConfigurationResponse (Maybe Text)

Required. The unique ID that Amazon MQ generates for the configuration.

describeConfigurationResponse_description :: Lens' DescribeConfigurationResponse (Maybe Text)

Required. The description of the configuration.

describeConfigurationResponse_engineType :: Lens' DescribeConfigurationResponse (Maybe EngineType)

Required. The type of broker engine. Currently, Amazon MQ supports ACTIVEMQ and RABBITMQ.

describeConfigurationResponse_tags :: Lens' DescribeConfigurationResponse (Maybe (HashMap Text Text))

The list of all tags associated with this configuration.

describeConfigurationResponse_httpStatus :: Lens' DescribeConfigurationResponse Int

The response's http status code.