Amazonka.MQ.ListConfigurationRevisions

Description

Returns a list of all revisions for the specified configuration.

Creating a Request

data ListConfigurationRevisions

See: newListConfigurationRevisions smart constructor.

Constructors

ListConfigurationRevisions'

Fields

nextToken :: Maybe Text

The token that specifies the next page of results Amazon MQ should return. To request the first page, leave nextToken empty.

maxResults :: Maybe Natural

The maximum number of brokers that Amazon MQ can return per page (20 by default). This value must be an integer from 5 to 100.

configurationId :: Text

The unique ID that Amazon MQ generates for the configuration.

newListConfigurationRevisions

Arguments

:: Text
$sel:configurationId:ListConfigurationRevisions'

-> ListConfigurationRevisions

Create a value of ListConfigurationRevisions with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:nextToken:ListConfigurationRevisions', listConfigurationRevisions_nextToken - The token that specifies the next page of results Amazon MQ should return. To request the first page, leave nextToken empty.

$sel:maxResults:ListConfigurationRevisions', listConfigurationRevisions_maxResults - The maximum number of brokers that Amazon MQ can return per page (20 by default). This value must be an integer from 5 to 100.

$sel:configurationId:ListConfigurationRevisions', listConfigurationRevisions_configurationId - The unique ID that Amazon MQ generates for the configuration.

Request Lenses

listConfigurationRevisions_nextToken :: Lens' ListConfigurationRevisions (Maybe Text)

The token that specifies the next page of results Amazon MQ should return. To request the first page, leave nextToken empty.

listConfigurationRevisions_maxResults :: Lens' ListConfigurationRevisions (Maybe Natural)

The maximum number of brokers that Amazon MQ can return per page (20 by default). This value must be an integer from 5 to 100.

listConfigurationRevisions_configurationId :: Lens' ListConfigurationRevisions Text

The unique ID that Amazon MQ generates for the configuration.

Destructuring the Response

data ListConfigurationRevisionsResponse

See: newListConfigurationRevisionsResponse smart constructor.

Constructors

ListConfigurationRevisionsResponse'

Fields

configurationId :: Maybe Text

The unique ID that Amazon MQ generates for the configuration.

nextToken :: Maybe Text

The token that specifies the next page of results Amazon MQ should return. To request the first page, leave nextToken empty.

revisions :: Maybe [ConfigurationRevision]

The list of all revisions for the specified configuration.

maxResults :: Maybe Int

The maximum number of configuration revisions that can be returned per page (20 by default). This value must be an integer from 5 to 100.

httpStatus :: Int

The response's http status code.

newListConfigurationRevisionsResponse

Arguments

:: Int
$sel:httpStatus:ListConfigurationRevisionsResponse'

-> ListConfigurationRevisionsResponse

Create a value of ListConfigurationRevisionsResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:configurationId:ListConfigurationRevisions', listConfigurationRevisionsResponse_configurationId - The unique ID that Amazon MQ generates for the configuration.

$sel:nextToken:ListConfigurationRevisions', listConfigurationRevisionsResponse_nextToken - The token that specifies the next page of results Amazon MQ should return. To request the first page, leave nextToken empty.

$sel:revisions:ListConfigurationRevisionsResponse', listConfigurationRevisionsResponse_revisions - The list of all revisions for the specified configuration.

$sel:maxResults:ListConfigurationRevisions', listConfigurationRevisionsResponse_maxResults - The maximum number of configuration revisions that can be returned per page (20 by default). This value must be an integer from 5 to 100.

$sel:httpStatus:ListConfigurationRevisionsResponse', listConfigurationRevisionsResponse_httpStatus - The response's http status code.

Response Lenses

listConfigurationRevisionsResponse_configurationId :: Lens' ListConfigurationRevisionsResponse (Maybe Text)

The unique ID that Amazon MQ generates for the configuration.

listConfigurationRevisionsResponse_nextToken :: Lens' ListConfigurationRevisionsResponse (Maybe Text)

The token that specifies the next page of results Amazon MQ should return. To request the first page, leave nextToken empty.

listConfigurationRevisionsResponse_revisions :: Lens' ListConfigurationRevisionsResponse (Maybe [ConfigurationRevision])

The list of all revisions for the specified configuration.

listConfigurationRevisionsResponse_maxResults :: Lens' ListConfigurationRevisionsResponse (Maybe Int)

The maximum number of configuration revisions that can be returned per page (20 by default). This value must be an integer from 5 to 100.

listConfigurationRevisionsResponse_httpStatus :: Lens' ListConfigurationRevisionsResponse Int

The response's http status code.