Amazonka.MQ.Types

Description

Service Configuration

defaultService :: Service

API version 2017-11-27 of the Amazon MQ SDK configuration.

Errors

_ConflictException :: AsError a => Getting (First ServiceError) a ServiceError

Returns information about an error.

_ForbiddenException :: AsError a => Getting (First ServiceError) a ServiceError

Returns information about an error.

_NotFoundException :: AsError a => Getting (First ServiceError) a ServiceError

Returns information about an error.

_InternalServerErrorException :: AsError a => Getting (First ServiceError) a ServiceError

Returns information about an error.

_UnauthorizedException :: AsError a => Getting (First ServiceError) a ServiceError

Returns information about an error.

_BadRequestException :: AsError a => Getting (First ServiceError) a ServiceError

Returns information about an error.

AuthenticationStrategy

newtype AuthenticationStrategy

Optional. The authentication strategy used to secure the broker. The default is SIMPLE.

Constructors

AuthenticationStrategy'

Fields

fromAuthenticationStrategy :: Text

Bundled Patterns

pattern AuthenticationStrategy_LDAP :: AuthenticationStrategy

pattern AuthenticationStrategy_SIMPLE :: AuthenticationStrategy

BrokerState

newtype BrokerState

The broker's status.

Constructors

BrokerState'

Fields

fromBrokerState :: Text

Bundled Patterns

pattern BrokerState_CREATION_FAILED :: BrokerState

pattern BrokerState_CREATION_IN_PROGRESS :: BrokerState

pattern BrokerState_DELETION_IN_PROGRESS :: BrokerState

pattern BrokerState_REBOOT_IN_PROGRESS :: BrokerState

pattern BrokerState_RUNNING :: BrokerState

BrokerStorageType

newtype BrokerStorageType

The broker's storage type.

EFS is not supported for RabbitMQ engine type.

Constructors

BrokerStorageType'

Fields

fromBrokerStorageType :: Text

Bundled Patterns

pattern BrokerStorageType_EBS :: BrokerStorageType

pattern BrokerStorageType_EFS :: BrokerStorageType

ChangeType

newtype ChangeType

The type of change pending for the ActiveMQ user.

Constructors

ChangeType'

Fields

fromChangeType :: Text

Bundled Patterns

pattern ChangeType_CREATE :: ChangeType

pattern ChangeType_DELETE :: ChangeType

pattern ChangeType_UPDATE :: ChangeType

DayOfWeek

newtype DayOfWeek

Constructors

DayOfWeek'

Fields

fromDayOfWeek :: Text

Bundled Patterns

pattern DayOfWeek_FRIDAY :: DayOfWeek

pattern DayOfWeek_MONDAY :: DayOfWeek

pattern DayOfWeek_SATURDAY :: DayOfWeek

pattern DayOfWeek_SUNDAY :: DayOfWeek

pattern DayOfWeek_THURSDAY :: DayOfWeek

pattern DayOfWeek_TUESDAY :: DayOfWeek

pattern DayOfWeek_WEDNESDAY :: DayOfWeek

DeploymentMode

newtype DeploymentMode

The broker's deployment mode.

Constructors

DeploymentMode'

Fields

fromDeploymentMode :: Text

Bundled Patterns

pattern DeploymentMode_ACTIVE_STANDBY_MULTI_AZ :: DeploymentMode

pattern DeploymentMode_CLUSTER_MULTI_AZ :: DeploymentMode

pattern DeploymentMode_SINGLE_INSTANCE :: DeploymentMode

EngineType

newtype EngineType

The type of broker engine. Amazon MQ supports ActiveMQ and RabbitMQ.

Constructors

EngineType'

Fields

fromEngineType :: Text

Bundled Patterns

pattern EngineType_ACTIVEMQ :: EngineType

pattern EngineType_RABBITMQ :: EngineType

SanitizationWarningReason

newtype SanitizationWarningReason

The reason for which the XML elements or attributes were sanitized.

Constructors

SanitizationWarningReason'

Fields

fromSanitizationWarningReason :: Text

Bundled Patterns

pattern SanitizationWarningReason_DISALLOWED_ATTRIBUTE_REMOVED :: SanitizationWarningReason

pattern SanitizationWarningReason_DISALLOWED_ELEMENT_REMOVED :: SanitizationWarningReason

pattern SanitizationWarningReason_INVALID_ATTRIBUTE_VALUE_REMOVED :: SanitizationWarningReason

AvailabilityZone

data AvailabilityZone

Name of the availability zone.

See: newAvailabilityZone smart constructor.

Constructors

AvailabilityZone'

Fields

name :: Maybe Text

Id for the availability zone.

newAvailabilityZone :: AvailabilityZone

Create a value of AvailabilityZone with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:name:AvailabilityZone', availabilityZone_name - Id for the availability zone.

availabilityZone_name :: Lens' AvailabilityZone (Maybe Text)

Id for the availability zone.

BrokerEngineType

data BrokerEngineType

Types of broker engines.

See: newBrokerEngineType smart constructor.

Constructors

BrokerEngineType'

Fields

engineVersions :: Maybe [EngineVersion]

The list of engine versions.

engineType :: Maybe EngineType

The broker's engine type.

newBrokerEngineType :: BrokerEngineType

Create a value of BrokerEngineType with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:engineVersions:BrokerEngineType', brokerEngineType_engineVersions - The list of engine versions.

$sel:engineType:BrokerEngineType', brokerEngineType_engineType - The broker's engine type.

brokerEngineType_engineVersions :: Lens' BrokerEngineType (Maybe [EngineVersion])

The list of engine versions.

brokerEngineType_engineType :: Lens' BrokerEngineType (Maybe EngineType)

The broker's engine type.

BrokerInstance

data BrokerInstance

Returns information about all brokers.

See: newBrokerInstance smart constructor.

Constructors

BrokerInstance'

Fields

ipAddress :: Maybe Text

The IP address of the Elastic Network Interface (ENI) attached to the broker. Does not apply to RabbitMQ brokers.

consoleURL :: Maybe Text

The brokers web console URL.

endpoints :: Maybe [Text]

The broker's wire-level protocol endpoints.

newBrokerInstance :: BrokerInstance

Create a value of BrokerInstance with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:ipAddress:BrokerInstance', brokerInstance_ipAddress - The IP address of the Elastic Network Interface (ENI) attached to the broker. Does not apply to RabbitMQ brokers.

$sel:consoleURL:BrokerInstance', brokerInstance_consoleURL - The brokers web console URL.

$sel:endpoints:BrokerInstance', brokerInstance_endpoints - The broker's wire-level protocol endpoints.

brokerInstance_ipAddress :: Lens' BrokerInstance (Maybe Text)

The IP address of the Elastic Network Interface (ENI) attached to the broker. Does not apply to RabbitMQ brokers.

brokerInstance_consoleURL :: Lens' BrokerInstance (Maybe Text)

The brokers web console URL.

brokerInstance_endpoints :: Lens' BrokerInstance (Maybe [Text])

The broker's wire-level protocol endpoints.

BrokerInstanceOption

data BrokerInstanceOption

Option for host instance type.

See: newBrokerInstanceOption smart constructor.

Constructors

BrokerInstanceOption'

Fields

supportedEngineVersions :: Maybe [Text]

The list of supported engine versions.

availabilityZones :: Maybe [AvailabilityZone]

The list of available az.

supportedDeploymentModes :: Maybe [DeploymentMode]

The list of supported deployment modes.

engineType :: Maybe EngineType

The broker's engine type.

hostInstanceType :: Maybe Text

The broker's instance type.

storageType :: Maybe BrokerStorageType

The broker's storage type.

newBrokerInstanceOption :: BrokerInstanceOption

Create a value of BrokerInstanceOption with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:supportedEngineVersions:BrokerInstanceOption', brokerInstanceOption_supportedEngineVersions - The list of supported engine versions.

$sel:availabilityZones:BrokerInstanceOption', brokerInstanceOption_availabilityZones - The list of available az.

$sel:supportedDeploymentModes:BrokerInstanceOption', brokerInstanceOption_supportedDeploymentModes - The list of supported deployment modes.

$sel:engineType:BrokerInstanceOption', brokerInstanceOption_engineType - The broker's engine type.

$sel:hostInstanceType:BrokerInstanceOption', brokerInstanceOption_hostInstanceType - The broker's instance type.

$sel:storageType:BrokerInstanceOption', brokerInstanceOption_storageType - The broker's storage type.

brokerInstanceOption_supportedEngineVersions :: Lens' BrokerInstanceOption (Maybe [Text])

The list of supported engine versions.

brokerInstanceOption_availabilityZones :: Lens' BrokerInstanceOption (Maybe [AvailabilityZone])

The list of available az.

brokerInstanceOption_supportedDeploymentModes :: Lens' BrokerInstanceOption (Maybe [DeploymentMode])

The list of supported deployment modes.

brokerInstanceOption_engineType :: Lens' BrokerInstanceOption (Maybe EngineType)

The broker's engine type.

brokerInstanceOption_hostInstanceType :: Lens' BrokerInstanceOption (Maybe Text)

The broker's instance type.

brokerInstanceOption_storageType :: Lens' BrokerInstanceOption (Maybe BrokerStorageType)

The broker's storage type.

BrokerSummary

data BrokerSummary

Returns information about all brokers.

See: newBrokerSummary smart constructor.

Constructors

BrokerSummary'

Fields

brokerName :: Maybe Text

The broker's name. This value is unique in your AWS account, 1-50 characters long, and containing only letters, numbers, dashes, and underscores, and must not contain white spaces, brackets, wildcard characters, or special characters.

brokerState :: Maybe BrokerState

The broker's status.

created :: Maybe POSIX

The time when the broker was created.

brokerId :: Maybe Text

The unique ID that Amazon MQ generates for the broker.

brokerArn :: Maybe Text

The broker's Amazon Resource Name (ARN).

hostInstanceType :: Maybe Text

The broker's instance type.

deploymentMode :: DeploymentMode

The broker's deployment mode.

engineType :: EngineType

The type of broker engine.

newBrokerSummary

Arguments

:: DeploymentMode
$sel:deploymentMode:BrokerSummary'

-> EngineType
$sel:engineType:BrokerSummary'

-> BrokerSummary

Create a value of BrokerSummary with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:brokerName:BrokerSummary', brokerSummary_brokerName - The broker's name. This value is unique in your AWS account, 1-50 characters long, and containing only letters, numbers, dashes, and underscores, and must not contain white spaces, brackets, wildcard characters, or special characters.

$sel:brokerState:BrokerSummary', brokerSummary_brokerState - The broker's status.

$sel:created:BrokerSummary', brokerSummary_created - The time when the broker was created.

$sel:brokerId:BrokerSummary', brokerSummary_brokerId - The unique ID that Amazon MQ generates for the broker.

$sel:brokerArn:BrokerSummary', brokerSummary_brokerArn - The broker's Amazon Resource Name (ARN).

$sel:hostInstanceType:BrokerSummary', brokerSummary_hostInstanceType - The broker's instance type.

$sel:deploymentMode:BrokerSummary', brokerSummary_deploymentMode - The broker's deployment mode.

$sel:engineType:BrokerSummary', brokerSummary_engineType - The type of broker engine.

brokerSummary_brokerName :: Lens' BrokerSummary (Maybe Text)

The broker's name. This value is unique in your AWS account, 1-50 characters long, and containing only letters, numbers, dashes, and underscores, and must not contain white spaces, brackets, wildcard characters, or special characters.

brokerSummary_brokerState :: Lens' BrokerSummary (Maybe BrokerState)

The broker's status.

brokerSummary_created :: Lens' BrokerSummary (Maybe UTCTime)

The time when the broker was created.

brokerSummary_brokerId :: Lens' BrokerSummary (Maybe Text)

The unique ID that Amazon MQ generates for the broker.

brokerSummary_brokerArn :: Lens' BrokerSummary (Maybe Text)

The broker's Amazon Resource Name (ARN).

brokerSummary_hostInstanceType :: Lens' BrokerSummary (Maybe Text)

The broker's instance type.

brokerSummary_deploymentMode :: Lens' BrokerSummary DeploymentMode

The broker's deployment mode.

brokerSummary_engineType :: Lens' BrokerSummary EngineType

The type of broker engine.

Configuration

data Configuration

Returns information about all configurations.

See: newConfiguration smart constructor.

Constructors

Configuration'

Fields

tags :: Maybe (HashMap Text Text)

The list of all tags associated with this configuration.

description :: Text

Required. The description of the configuration.

engineVersion :: Text

Required. The broker engine's version. For a list of supported engine versions, see, Supported engines.

latestRevision :: ConfigurationRevision

Required. The latest revision of the configuration.

authenticationStrategy :: AuthenticationStrategy

Optional. The authentication strategy associated with the configuration. The default is SIMPLE.

engineType :: EngineType

Required. The type of broker engine. Currently, Amazon MQ supports ACTIVEMQ and RABBITMQ.

id :: Text

Required. The unique ID that Amazon MQ generates for the configuration.

arn :: Text

Required. The ARN of the configuration.

name :: Text

Required. The name of the configuration. This value can contain only alphanumeric characters, dashes, periods, underscores, and tildes (- . _ ~). This value must be 1-150 characters long.

created :: POSIX

Required. The date and time of the configuration revision.

newConfiguration

Arguments

:: Text
$sel:description:Configuration'

-> Text
$sel:engineVersion:Configuration'

-> ConfigurationRevision
$sel:latestRevision:Configuration'

-> AuthenticationStrategy
$sel:authenticationStrategy:Configuration'

-> EngineType
$sel:engineType:Configuration'

-> Text
$sel:id:Configuration'

-> Text
$sel:arn:Configuration'

-> Text
$sel:name:Configuration'

-> UTCTime
$sel:created:Configuration'

-> Configuration

Create a value of Configuration with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:tags:Configuration', configuration_tags - The list of all tags associated with this configuration.

$sel:description:Configuration', configuration_description - Required. The description of the configuration.

$sel:engineVersion:Configuration', configuration_engineVersion - Required. The broker engine's version. For a list of supported engine versions, see, Supported engines.

$sel:latestRevision:Configuration', configuration_latestRevision - Required. The latest revision of the configuration.

$sel:authenticationStrategy:Configuration', configuration_authenticationStrategy - Optional. The authentication strategy associated with the configuration. The default is SIMPLE.

$sel:engineType:Configuration', configuration_engineType - Required. The type of broker engine. Currently, Amazon MQ supports ACTIVEMQ and RABBITMQ.

$sel:id:Configuration', configuration_id - Required. The unique ID that Amazon MQ generates for the configuration.

$sel:arn:Configuration', configuration_arn - Required. The ARN of the configuration.

$sel:name:Configuration', configuration_name - Required. The name of the configuration. This value can contain only alphanumeric characters, dashes, periods, underscores, and tildes (- . _ ~). This value must be 1-150 characters long.

$sel:created:Configuration', configuration_created - Required. The date and time of the configuration revision.

configuration_tags :: Lens' Configuration (Maybe (HashMap Text Text))

The list of all tags associated with this configuration.

configuration_description :: Lens' Configuration Text

Required. The description of the configuration.

configuration_engineVersion :: Lens' Configuration Text

Required. The broker engine's version. For a list of supported engine versions, see, Supported engines.

configuration_latestRevision :: Lens' Configuration ConfigurationRevision

Required. The latest revision of the configuration.

configuration_authenticationStrategy :: Lens' Configuration AuthenticationStrategy

Optional. The authentication strategy associated with the configuration. The default is SIMPLE.

configuration_engineType :: Lens' Configuration EngineType

Required. The type of broker engine. Currently, Amazon MQ supports ACTIVEMQ and RABBITMQ.

configuration_id :: Lens' Configuration Text

Required. The unique ID that Amazon MQ generates for the configuration.

configuration_arn :: Lens' Configuration Text

Required. The ARN of the configuration.

configuration_name :: Lens' Configuration Text

Required. The name of the configuration. This value can contain only alphanumeric characters, dashes, periods, underscores, and tildes (- . _ ~). This value must be 1-150 characters long.

configuration_created :: Lens' Configuration UTCTime

Required. The date and time of the configuration revision.

ConfigurationId

data ConfigurationId

A list of information about the configuration.

Does not apply to RabbitMQ brokers.

See: newConfigurationId smart constructor.

Constructors

ConfigurationId'

Fields

revision :: Maybe Int

The revision number of the configuration.

id :: Text

Required. The unique ID that Amazon MQ generates for the configuration.

newConfigurationId

Arguments

:: Text
$sel:id:ConfigurationId'

-> ConfigurationId

Create a value of ConfigurationId with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:revision:ConfigurationId', configurationId_revision - The revision number of the configuration.

$sel:id:ConfigurationId', configurationId_id - Required. The unique ID that Amazon MQ generates for the configuration.

configurationId_revision :: Lens' ConfigurationId (Maybe Int)

The revision number of the configuration.

configurationId_id :: Lens' ConfigurationId Text

Required. The unique ID that Amazon MQ generates for the configuration.

ConfigurationRevision

data ConfigurationRevision

Returns information about the specified configuration revision.

See: newConfigurationRevision smart constructor.

Constructors

ConfigurationRevision'

Fields

description :: Maybe Text

The description of the configuration revision.

revision :: Int

Required. The revision number of the configuration.

created :: POSIX

Required. The date and time of the configuration revision.

newConfigurationRevision

Arguments

:: Int
$sel:revision:ConfigurationRevision'

-> UTCTime
$sel:created:ConfigurationRevision'

-> ConfigurationRevision

Create a value of ConfigurationRevision with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:description:ConfigurationRevision', configurationRevision_description - The description of the configuration revision.

$sel:revision:ConfigurationRevision', configurationRevision_revision - Required. The revision number of the configuration.

$sel:created:ConfigurationRevision', configurationRevision_created - Required. The date and time of the configuration revision.

configurationRevision_description :: Lens' ConfigurationRevision (Maybe Text)

The description of the configuration revision.

configurationRevision_revision :: Lens' ConfigurationRevision Int

Required. The revision number of the configuration.

configurationRevision_created :: Lens' ConfigurationRevision UTCTime

Required. The date and time of the configuration revision.

Configurations

data Configurations

Broker configuration information

See: newConfigurations smart constructor.

Constructors

Configurations'

Fields

pending :: Maybe ConfigurationId

The broker's pending configuration.

history :: Maybe [ConfigurationId]

The history of configurations applied to the broker.

current :: Maybe ConfigurationId

The broker's current configuration.

newConfigurations :: Configurations

Create a value of Configurations with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:pending:Configurations', configurations_pending - The broker's pending configuration.

$sel:history:Configurations', configurations_history - The history of configurations applied to the broker.

$sel:current:Configurations', configurations_current - The broker's current configuration.

configurations_pending :: Lens' Configurations (Maybe ConfigurationId)

The broker's pending configuration.

configurations_history :: Lens' Configurations (Maybe [ConfigurationId])

The history of configurations applied to the broker.

configurations_current :: Lens' Configurations (Maybe ConfigurationId)

The broker's current configuration.

EncryptionOptions

data EncryptionOptions

Does not apply to RabbitMQ brokers.

Encryption options for the broker.

See: newEncryptionOptions smart constructor.

Constructors

EncryptionOptions'

Fields

kmsKeyId :: Maybe Text

The customer master key (CMK) to use for the AWS Key Management Service (KMS). This key is used to encrypt your data at rest. If not provided, Amazon MQ will use a default CMK to encrypt your data.

useAwsOwnedKey :: Bool

Enables the use of an AWS owned CMK using AWS Key Management Service (KMS). Set to true by default, if no value is provided, for example, for RabbitMQ brokers.

newEncryptionOptions

Arguments

:: Bool
$sel:useAwsOwnedKey:EncryptionOptions'

-> EncryptionOptions

Create a value of EncryptionOptions with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:kmsKeyId:EncryptionOptions', encryptionOptions_kmsKeyId - The customer master key (CMK) to use for the AWS Key Management Service (KMS). This key is used to encrypt your data at rest. If not provided, Amazon MQ will use a default CMK to encrypt your data.

$sel:useAwsOwnedKey:EncryptionOptions', encryptionOptions_useAwsOwnedKey - Enables the use of an AWS owned CMK using AWS Key Management Service (KMS). Set to true by default, if no value is provided, for example, for RabbitMQ brokers.

encryptionOptions_kmsKeyId :: Lens' EncryptionOptions (Maybe Text)

The customer master key (CMK) to use for the AWS Key Management Service (KMS). This key is used to encrypt your data at rest. If not provided, Amazon MQ will use a default CMK to encrypt your data.

encryptionOptions_useAwsOwnedKey :: Lens' EncryptionOptions Bool

Enables the use of an AWS owned CMK using AWS Key Management Service (KMS). Set to true by default, if no value is provided, for example, for RabbitMQ brokers.

EngineVersion

data EngineVersion

Id of the engine version.

See: newEngineVersion smart constructor.

Constructors

EngineVersion'

Fields

name :: Maybe Text

Id for the version.

newEngineVersion :: EngineVersion

Create a value of EngineVersion with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:name:EngineVersion', engineVersion_name - Id for the version.

engineVersion_name :: Lens' EngineVersion (Maybe Text)

Id for the version.

LdapServerMetadataInput

data LdapServerMetadataInput

Optional. The metadata of the LDAP server used to authenticate and authorize connections to the broker.

Does not apply to RabbitMQ brokers.

See: newLdapServerMetadataInput smart constructor.

Constructors

LdapServerMetadataInput'

Fields

userRoleName :: Maybe Text

Specifies the name of the LDAP attribute for the user group membership.

userSearchSubtree :: Maybe Bool

The directory search scope for the user. If set to true, scope is to search the entire subtree.

roleSearchSubtree :: Maybe Bool

The directory search scope for the role. If set to true, scope is to search the entire subtree.

roleName :: Maybe Text

Specifies the LDAP attribute that identifies the group name attribute in the object returned from the group membership query.

hosts :: [Text]

Specifies the location of the LDAP server such as AWS Directory Service for Microsoft Active Directory . Optional failover server.

userSearchMatching :: Text

The LDAP search filter used to find users within the userBase. The client's username is substituted into the {0} placeholder in the search filter. For example, if this option is set to (uid={0}) and the received username is janedoe, the search filter becomes (uid=janedoe) after string substitution. It will result in matching an entry like uid=janedoe, ou=Users,ou=corp, dc=corp, dc=example, dc=com.

userBase :: Text

Select a particular subtree of the directory information tree (DIT) to search for user entries. The subtree is specified by a DN, which specifies the base node of the subtree. For example, by setting this option to ou=Users,ou=corp, dc=corp, dc=example, dc=com, the search for user entries is restricted to the subtree beneath ou=Users, ou=corp, dc=corp, dc=example, dc=com.

roleSearchMatching :: Text

The LDAP search filter used to find roles within the roleBase. The distinguished name of the user matched by userSearchMatching is substituted into the {0} placeholder in the search filter. The client's username is substituted into the {1} placeholder. For example, if you set this option to (member=uid={1})for the user janedoe, the search filter becomes (member=uid=janedoe) after string substitution. It matches all role entries that have a member attribute equal to uid=janedoe under the subtree selected by the roleBase.

serviceAccountUsername :: Text

Service account username. A service account is an account in your LDAP server that has access to initiate a connection. For example, cn=admin,dc=corp, dc=example, dc=com.

roleBase :: Text

The distinguished name of the node in the directory information tree (DIT) to search for roles or groups. For example, ou=group, ou=corp, dc=corp, dc=example, dc=com.

serviceAccountPassword :: Text

Service account password. A service account is an account in your LDAP server that has access to initiate a connection. For example, cn=admin,dc=corp, dc=example, dc=com.

newLdapServerMetadataInput

Arguments

:: Text
$sel:userSearchMatching:LdapServerMetadataInput'

-> Text
$sel:userBase:LdapServerMetadataInput'

-> Text
$sel:roleSearchMatching:LdapServerMetadataInput'

-> Text
$sel:serviceAccountUsername:LdapServerMetadataInput'

-> Text
$sel:roleBase:LdapServerMetadataInput'

-> Text
$sel:serviceAccountPassword:LdapServerMetadataInput'

-> LdapServerMetadataInput

Create a value of LdapServerMetadataInput with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:userRoleName:LdapServerMetadataInput', ldapServerMetadataInput_userRoleName - Specifies the name of the LDAP attribute for the user group membership.

$sel:userSearchSubtree:LdapServerMetadataInput', ldapServerMetadataInput_userSearchSubtree - The directory search scope for the user. If set to true, scope is to search the entire subtree.

$sel:roleSearchSubtree:LdapServerMetadataInput', ldapServerMetadataInput_roleSearchSubtree - The directory search scope for the role. If set to true, scope is to search the entire subtree.

$sel:roleName:LdapServerMetadataInput', ldapServerMetadataInput_roleName - Specifies the LDAP attribute that identifies the group name attribute in the object returned from the group membership query.

$sel:hosts:LdapServerMetadataInput', ldapServerMetadataInput_hosts - Specifies the location of the LDAP server such as AWS Directory Service for Microsoft Active Directory . Optional failover server.

$sel:userSearchMatching:LdapServerMetadataInput', ldapServerMetadataInput_userSearchMatching - The LDAP search filter used to find users within the userBase. The client's username is substituted into the {0} placeholder in the search filter. For example, if this option is set to (uid={0}) and the received username is janedoe, the search filter becomes (uid=janedoe) after string substitution. It will result in matching an entry like uid=janedoe, ou=Users,ou=corp, dc=corp, dc=example, dc=com.

$sel:userBase:LdapServerMetadataInput', ldapServerMetadataInput_userBase - Select a particular subtree of the directory information tree (DIT) to search for user entries. The subtree is specified by a DN, which specifies the base node of the subtree. For example, by setting this option to ou=Users,ou=corp, dc=corp, dc=example, dc=com, the search for user entries is restricted to the subtree beneath ou=Users, ou=corp, dc=corp, dc=example, dc=com.

$sel:roleSearchMatching:LdapServerMetadataInput', ldapServerMetadataInput_roleSearchMatching - The LDAP search filter used to find roles within the roleBase. The distinguished name of the user matched by userSearchMatching is substituted into the {0} placeholder in the search filter. The client's username is substituted into the {1} placeholder. For example, if you set this option to (member=uid={1})for the user janedoe, the search filter becomes (member=uid=janedoe) after string substitution. It matches all role entries that have a member attribute equal to uid=janedoe under the subtree selected by the roleBase.

$sel:serviceAccountUsername:LdapServerMetadataInput', ldapServerMetadataInput_serviceAccountUsername - Service account username. A service account is an account in your LDAP server that has access to initiate a connection. For example, cn=admin,dc=corp, dc=example, dc=com.

$sel:roleBase:LdapServerMetadataInput', ldapServerMetadataInput_roleBase - The distinguished name of the node in the directory information tree (DIT) to search for roles or groups. For example, ou=group, ou=corp, dc=corp, dc=example, dc=com.

$sel:serviceAccountPassword:LdapServerMetadataInput', ldapServerMetadataInput_serviceAccountPassword - Service account password. A service account is an account in your LDAP server that has access to initiate a connection. For example, cn=admin,dc=corp, dc=example, dc=com.

ldapServerMetadataInput_userRoleName :: Lens' LdapServerMetadataInput (Maybe Text)

Specifies the name of the LDAP attribute for the user group membership.

ldapServerMetadataInput_userSearchSubtree :: Lens' LdapServerMetadataInput (Maybe Bool)

The directory search scope for the user. If set to true, scope is to search the entire subtree.

ldapServerMetadataInput_roleSearchSubtree :: Lens' LdapServerMetadataInput (Maybe Bool)

The directory search scope for the role. If set to true, scope is to search the entire subtree.

ldapServerMetadataInput_roleName :: Lens' LdapServerMetadataInput (Maybe Text)

Specifies the LDAP attribute that identifies the group name attribute in the object returned from the group membership query.

ldapServerMetadataInput_hosts :: Lens' LdapServerMetadataInput [Text]

Specifies the location of the LDAP server such as AWS Directory Service for Microsoft Active Directory . Optional failover server.

ldapServerMetadataInput_userSearchMatching :: Lens' LdapServerMetadataInput Text

The LDAP search filter used to find users within the userBase. The client's username is substituted into the {0} placeholder in the search filter. For example, if this option is set to (uid={0}) and the received username is janedoe, the search filter becomes (uid=janedoe) after string substitution. It will result in matching an entry like uid=janedoe, ou=Users,ou=corp, dc=corp, dc=example, dc=com.

ldapServerMetadataInput_userBase :: Lens' LdapServerMetadataInput Text

Select a particular subtree of the directory information tree (DIT) to search for user entries. The subtree is specified by a DN, which specifies the base node of the subtree. For example, by setting this option to ou=Users,ou=corp, dc=corp, dc=example, dc=com, the search for user entries is restricted to the subtree beneath ou=Users, ou=corp, dc=corp, dc=example, dc=com.

ldapServerMetadataInput_roleSearchMatching :: Lens' LdapServerMetadataInput Text

The LDAP search filter used to find roles within the roleBase. The distinguished name of the user matched by userSearchMatching is substituted into the {0} placeholder in the search filter. The client's username is substituted into the {1} placeholder. For example, if you set this option to (member=uid={1})for the user janedoe, the search filter becomes (member=uid=janedoe) after string substitution. It matches all role entries that have a member attribute equal to uid=janedoe under the subtree selected by the roleBase.

ldapServerMetadataInput_serviceAccountUsername :: Lens' LdapServerMetadataInput Text

Service account username. A service account is an account in your LDAP server that has access to initiate a connection. For example, cn=admin,dc=corp, dc=example, dc=com.

ldapServerMetadataInput_roleBase :: Lens' LdapServerMetadataInput Text

The distinguished name of the node in the directory information tree (DIT) to search for roles or groups. For example, ou=group, ou=corp, dc=corp, dc=example, dc=com.

ldapServerMetadataInput_serviceAccountPassword :: Lens' LdapServerMetadataInput Text

Service account password. A service account is an account in your LDAP server that has access to initiate a connection. For example, cn=admin,dc=corp, dc=example, dc=com.

LdapServerMetadataOutput

data LdapServerMetadataOutput

Optional. The metadata of the LDAP server used to authenticate and authorize connections to the broker.

See: newLdapServerMetadataOutput smart constructor.

Constructors

LdapServerMetadataOutput'

Fields

userRoleName :: Maybe Text

Specifies the name of the LDAP attribute for the user group membership.

userSearchSubtree :: Maybe Bool

The directory search scope for the user. If set to true, scope is to search the entire subtree.

roleSearchSubtree :: Maybe Bool

The directory search scope for the role. If set to true, scope is to search the entire subtree.

roleName :: Maybe Text

Specifies the LDAP attribute that identifies the group name attribute in the object returned from the group membership query.

hosts :: [Text]

Specifies the location of the LDAP server such as AWS Directory Service for Microsoft Active Directory . Optional failover server.

userSearchMatching :: Text

The LDAP search filter used to find users within the userBase. The client's username is substituted into the {0} placeholder in the search filter. For example, if this option is set to (uid={0}) and the received username is janedoe, the search filter becomes (uid=janedoe) after string substitution. It will result in matching an entry like uid=janedoe, ou=Users,ou=corp, dc=corp, dc=example, dc=com.

userBase :: Text

Select a particular subtree of the directory information tree (DIT) to search for user entries. The subtree is specified by a DN, which specifies the base node of the subtree. For example, by setting this option to ou=Users,ou=corp, dc=corp, dc=example, dc=com, the search for user entries is restricted to the subtree beneath ou=Users, ou=corp, dc=corp, dc=example, dc=com.

roleSearchMatching :: Text

The LDAP search filter used to find roles within the roleBase. The distinguished name of the user matched by userSearchMatching is substituted into the {0} placeholder in the search filter. The client's username is substituted into the {1} placeholder. For example, if you set this option to (member=uid={1})for the user janedoe, the search filter becomes (member=uid=janedoe) after string substitution. It matches all role entries that have a member attribute equal to uid=janedoe under the subtree selected by the roleBase.

serviceAccountUsername :: Text

Service account username. A service account is an account in your LDAP server that has access to initiate a connection. For example, cn=admin,dc=corp, dc=example, dc=com.

roleBase :: Text

The distinguished name of the node in the directory information tree (DIT) to search for roles or groups. For example, ou=group, ou=corp, dc=corp, dc=example, dc=com.

newLdapServerMetadataOutput

Arguments

:: Text
$sel:userSearchMatching:LdapServerMetadataOutput'

-> Text
$sel:userBase:LdapServerMetadataOutput'

-> Text
$sel:roleSearchMatching:LdapServerMetadataOutput'

-> Text
$sel:serviceAccountUsername:LdapServerMetadataOutput'

-> Text
$sel:roleBase:LdapServerMetadataOutput'

-> LdapServerMetadataOutput

Create a value of LdapServerMetadataOutput with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:userRoleName:LdapServerMetadataOutput', ldapServerMetadataOutput_userRoleName - Specifies the name of the LDAP attribute for the user group membership.

$sel:userSearchSubtree:LdapServerMetadataOutput', ldapServerMetadataOutput_userSearchSubtree - The directory search scope for the user. If set to true, scope is to search the entire subtree.

$sel:roleSearchSubtree:LdapServerMetadataOutput', ldapServerMetadataOutput_roleSearchSubtree - The directory search scope for the role. If set to true, scope is to search the entire subtree.

$sel:roleName:LdapServerMetadataOutput', ldapServerMetadataOutput_roleName - Specifies the LDAP attribute that identifies the group name attribute in the object returned from the group membership query.

$sel:hosts:LdapServerMetadataOutput', ldapServerMetadataOutput_hosts - Specifies the location of the LDAP server such as AWS Directory Service for Microsoft Active Directory . Optional failover server.

$sel:userSearchMatching:LdapServerMetadataOutput', ldapServerMetadataOutput_userSearchMatching - The LDAP search filter used to find users within the userBase. The client's username is substituted into the {0} placeholder in the search filter. For example, if this option is set to (uid={0}) and the received username is janedoe, the search filter becomes (uid=janedoe) after string substitution. It will result in matching an entry like uid=janedoe, ou=Users,ou=corp, dc=corp, dc=example, dc=com.

$sel:userBase:LdapServerMetadataOutput', ldapServerMetadataOutput_userBase - Select a particular subtree of the directory information tree (DIT) to search for user entries. The subtree is specified by a DN, which specifies the base node of the subtree. For example, by setting this option to ou=Users,ou=corp, dc=corp, dc=example, dc=com, the search for user entries is restricted to the subtree beneath ou=Users, ou=corp, dc=corp, dc=example, dc=com.

$sel:roleSearchMatching:LdapServerMetadataOutput', ldapServerMetadataOutput_roleSearchMatching - The LDAP search filter used to find roles within the roleBase. The distinguished name of the user matched by userSearchMatching is substituted into the {0} placeholder in the search filter. The client's username is substituted into the {1} placeholder. For example, if you set this option to (member=uid={1})for the user janedoe, the search filter becomes (member=uid=janedoe) after string substitution. It matches all role entries that have a member attribute equal to uid=janedoe under the subtree selected by the roleBase.

$sel:serviceAccountUsername:LdapServerMetadataOutput', ldapServerMetadataOutput_serviceAccountUsername - Service account username. A service account is an account in your LDAP server that has access to initiate a connection. For example, cn=admin,dc=corp, dc=example, dc=com.

$sel:roleBase:LdapServerMetadataOutput', ldapServerMetadataOutput_roleBase - The distinguished name of the node in the directory information tree (DIT) to search for roles or groups. For example, ou=group, ou=corp, dc=corp, dc=example, dc=com.

ldapServerMetadataOutput_userRoleName :: Lens' LdapServerMetadataOutput (Maybe Text)

Specifies the name of the LDAP attribute for the user group membership.

ldapServerMetadataOutput_userSearchSubtree :: Lens' LdapServerMetadataOutput (Maybe Bool)

The directory search scope for the user. If set to true, scope is to search the entire subtree.

ldapServerMetadataOutput_roleSearchSubtree :: Lens' LdapServerMetadataOutput (Maybe Bool)

The directory search scope for the role. If set to true, scope is to search the entire subtree.

ldapServerMetadataOutput_roleName :: Lens' LdapServerMetadataOutput (Maybe Text)

Specifies the LDAP attribute that identifies the group name attribute in the object returned from the group membership query.

ldapServerMetadataOutput_hosts :: Lens' LdapServerMetadataOutput [Text]

Specifies the location of the LDAP server such as AWS Directory Service for Microsoft Active Directory . Optional failover server.

ldapServerMetadataOutput_userSearchMatching :: Lens' LdapServerMetadataOutput Text

The LDAP search filter used to find users within the userBase. The client's username is substituted into the {0} placeholder in the search filter. For example, if this option is set to (uid={0}) and the received username is janedoe, the search filter becomes (uid=janedoe) after string substitution. It will result in matching an entry like uid=janedoe, ou=Users,ou=corp, dc=corp, dc=example, dc=com.

ldapServerMetadataOutput_userBase :: Lens' LdapServerMetadataOutput Text

Select a particular subtree of the directory information tree (DIT) to search for user entries. The subtree is specified by a DN, which specifies the base node of the subtree. For example, by setting this option to ou=Users,ou=corp, dc=corp, dc=example, dc=com, the search for user entries is restricted to the subtree beneath ou=Users, ou=corp, dc=corp, dc=example, dc=com.

ldapServerMetadataOutput_roleSearchMatching :: Lens' LdapServerMetadataOutput Text

The LDAP search filter used to find roles within the roleBase. The distinguished name of the user matched by userSearchMatching is substituted into the {0} placeholder in the search filter. The client's username is substituted into the {1} placeholder. For example, if you set this option to (member=uid={1})for the user janedoe, the search filter becomes (member=uid=janedoe) after string substitution. It matches all role entries that have a member attribute equal to uid=janedoe under the subtree selected by the roleBase.

ldapServerMetadataOutput_serviceAccountUsername :: Lens' LdapServerMetadataOutput Text

Service account username. A service account is an account in your LDAP server that has access to initiate a connection. For example, cn=admin,dc=corp, dc=example, dc=com.

ldapServerMetadataOutput_roleBase :: Lens' LdapServerMetadataOutput Text

The distinguished name of the node in the directory information tree (DIT) to search for roles or groups. For example, ou=group, ou=corp, dc=corp, dc=example, dc=com.

Logs

data Logs

The list of information about logs to be enabled for the specified broker.

See: newLogs smart constructor.

Constructors

Logs'

Fields

audit :: Maybe Bool

Enables audit logging. Every user management action made using JMX or the ActiveMQ Web Console is logged. Does not apply to RabbitMQ brokers.

general :: Maybe Bool

Enables general logging.

newLogs :: Logs

Create a value of Logs with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:audit:Logs', logs_audit - Enables audit logging. Every user management action made using JMX or the ActiveMQ Web Console is logged. Does not apply to RabbitMQ brokers.

$sel:general:Logs', logs_general - Enables general logging.

logs_audit :: Lens' Logs (Maybe Bool)

Enables audit logging. Every user management action made using JMX or the ActiveMQ Web Console is logged. Does not apply to RabbitMQ brokers.

logs_general :: Lens' Logs (Maybe Bool)

Enables general logging.

LogsSummary

data LogsSummary

The list of information about logs currently enabled and pending to be deployed for the specified broker.

See: newLogsSummary smart constructor.

Constructors

LogsSummary'

Fields

pending :: Maybe PendingLogs

The list of information about logs pending to be deployed for the specified broker.

audit :: Maybe Bool

Enables audit logging. Every user management action made using JMX or the ActiveMQ Web Console is logged.

auditLogGroup :: Maybe Text

The location of the CloudWatch Logs log group where audit logs are sent.

generalLogGroup :: Text

The location of the CloudWatch Logs log group where general logs are sent.

general :: Bool

Enables general logging.

newLogsSummary

Arguments

:: Text
$sel:generalLogGroup:LogsSummary'

-> Bool
$sel:general:LogsSummary'

-> LogsSummary

Create a value of LogsSummary with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:pending:LogsSummary', logsSummary_pending - The list of information about logs pending to be deployed for the specified broker.

$sel:audit:LogsSummary', logsSummary_audit - Enables audit logging. Every user management action made using JMX or the ActiveMQ Web Console is logged.

$sel:auditLogGroup:LogsSummary', logsSummary_auditLogGroup - The location of the CloudWatch Logs log group where audit logs are sent.

$sel:generalLogGroup:LogsSummary', logsSummary_generalLogGroup - The location of the CloudWatch Logs log group where general logs are sent.

$sel:general:LogsSummary', logsSummary_general - Enables general logging.

logsSummary_pending :: Lens' LogsSummary (Maybe PendingLogs)

The list of information about logs pending to be deployed for the specified broker.

logsSummary_audit :: Lens' LogsSummary (Maybe Bool)

Enables audit logging. Every user management action made using JMX or the ActiveMQ Web Console is logged.

logsSummary_auditLogGroup :: Lens' LogsSummary (Maybe Text)

The location of the CloudWatch Logs log group where audit logs are sent.

logsSummary_generalLogGroup :: Lens' LogsSummary Text

The location of the CloudWatch Logs log group where general logs are sent.

logsSummary_general :: Lens' LogsSummary Bool

Enables general logging.

PendingLogs

data PendingLogs

The list of information about logs to be enabled for the specified broker.

See: newPendingLogs smart constructor.

Constructors

PendingLogs'

Fields

audit :: Maybe Bool

Enables audit logging. Every user management action made using JMX or the ActiveMQ Web Console is logged.

general :: Maybe Bool

Enables general logging.

newPendingLogs :: PendingLogs

Create a value of PendingLogs with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:audit:PendingLogs', pendingLogs_audit - Enables audit logging. Every user management action made using JMX or the ActiveMQ Web Console is logged.

$sel:general:PendingLogs', pendingLogs_general - Enables general logging.

pendingLogs_audit :: Lens' PendingLogs (Maybe Bool)

Enables audit logging. Every user management action made using JMX or the ActiveMQ Web Console is logged.

pendingLogs_general :: Lens' PendingLogs (Maybe Bool)

Enables general logging.

SanitizationWarning

data SanitizationWarning

Returns information about the XML element or attribute that was sanitized in the configuration.

See: newSanitizationWarning smart constructor.

Constructors

SanitizationWarning'

Fields

attributeName :: Maybe Text

The name of the XML attribute that has been sanitized.

elementName :: Maybe Text

The name of the XML element that has been sanitized.

reason :: SanitizationWarningReason

Required. The reason for which the XML elements or attributes were sanitized.

newSanitizationWarning

Arguments

:: SanitizationWarningReason
$sel:reason:SanitizationWarning'

-> SanitizationWarning

Create a value of SanitizationWarning with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:attributeName:SanitizationWarning', sanitizationWarning_attributeName - The name of the XML attribute that has been sanitized.

$sel:elementName:SanitizationWarning', sanitizationWarning_elementName - The name of the XML element that has been sanitized.

$sel:reason:SanitizationWarning', sanitizationWarning_reason - Required. The reason for which the XML elements or attributes were sanitized.

sanitizationWarning_attributeName :: Lens' SanitizationWarning (Maybe Text)

The name of the XML attribute that has been sanitized.

sanitizationWarning_elementName :: Lens' SanitizationWarning (Maybe Text)

The name of the XML element that has been sanitized.

sanitizationWarning_reason :: Lens' SanitizationWarning SanitizationWarningReason

Required. The reason for which the XML elements or attributes were sanitized.

User

data User

A user associated with the broker. For RabbitMQ brokers, one and only one administrative user is accepted and created when a broker is first provisioned. All subsequent broker users are created by making RabbitMQ API calls directly to brokers or via the RabbitMQ web console.

See: newUser smart constructor.

Constructors

User'

Fields

groups :: Maybe [Text]

The list of groups (20 maximum) to which the ActiveMQ user belongs. This value can contain only alphanumeric characters, dashes, periods, underscores, and tildes (- . _ ~). This value must be 2-100 characters long. Does not apply to RabbitMQ brokers.

consoleAccess :: Maybe Bool

Enables access to the ActiveMQ Web Console for the ActiveMQ user. Does not apply to RabbitMQ brokers.

username :: Text

important>

Amazon MQ for ActiveMQ

For ActiveMQ brokers, this value can contain only alphanumeric characters, dashes, periods, underscores, and tildes (- . _ ~). This value must be 2-100 characters long.

/important>

Amazon MQ for RabbitMQ

For RabbitMQ brokers, this value can contain only alphanumeric characters, dashes, periods, underscores (- . _). This value must not contain a tilde (~) character. Amazon MQ prohibts using guest as a valid usename. This value must be 2-100 characters long.

password :: Text

Required. The password of the user. This value must be at least 12 characters long, must contain at least 4 unique characters, and must not contain commas, colons, or equal signs (,:=).

newUser

Arguments

:: Text
$sel:username:User'

-> Text
$sel:password:User'

-> User

Create a value of User with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:groups:User', user_groups - The list of groups (20 maximum) to which the ActiveMQ user belongs. This value can contain only alphanumeric characters, dashes, periods, underscores, and tildes (- . _ ~). This value must be 2-100 characters long. Does not apply to RabbitMQ brokers.

$sel:consoleAccess:User', user_consoleAccess - Enables access to the ActiveMQ Web Console for the ActiveMQ user. Does not apply to RabbitMQ brokers.

$sel:username:User', user_username - important>

Amazon MQ for ActiveMQ

For ActiveMQ brokers, this value can contain only alphanumeric characters, dashes, periods, underscores, and tildes (- . _ ~). This value must be 2-100 characters long.

/important>

Amazon MQ for RabbitMQ

For RabbitMQ brokers, this value can contain only alphanumeric characters, dashes, periods, underscores (- . _). This value must not contain a tilde (~) character. Amazon MQ prohibts using guest as a valid usename. This value must be 2-100 characters long.

$sel:password:User', user_password - Required. The password of the user. This value must be at least 12 characters long, must contain at least 4 unique characters, and must not contain commas, colons, or equal signs (,:=).

user_groups :: Lens' User (Maybe [Text])

The list of groups (20 maximum) to which the ActiveMQ user belongs. This value can contain only alphanumeric characters, dashes, periods, underscores, and tildes (- . _ ~). This value must be 2-100 characters long. Does not apply to RabbitMQ brokers.

user_consoleAccess :: Lens' User (Maybe Bool)

Enables access to the ActiveMQ Web Console for the ActiveMQ user. Does not apply to RabbitMQ brokers.

user_username :: Lens' User Text

important>

Amazon MQ for ActiveMQ

For ActiveMQ brokers, this value can contain only alphanumeric characters, dashes, periods, underscores, and tildes (- . _ ~). This value must be 2-100 characters long.

/important>

Amazon MQ for RabbitMQ

For RabbitMQ brokers, this value can contain only alphanumeric characters, dashes, periods, underscores (- . _). This value must not contain a tilde (~) character. Amazon MQ prohibts using guest as a valid usename. This value must be 2-100 characters long.

user_password :: Lens' User Text

Required. The password of the user. This value must be at least 12 characters long, must contain at least 4 unique characters, and must not contain commas, colons, or equal signs (,:=).

UserPendingChanges

data UserPendingChanges

Returns information about the status of the changes pending for the ActiveMQ user.

See: newUserPendingChanges smart constructor.

Constructors

UserPendingChanges'

Fields

groups :: Maybe [Text]

The list of groups (20 maximum) to which the ActiveMQ user belongs. This value can contain only alphanumeric characters, dashes, periods, underscores, and tildes (- . _ ~). This value must be 2-100 characters long.

consoleAccess :: Maybe Bool

Enables access to the the ActiveMQ Web Console for the ActiveMQ user.

pendingChange :: ChangeType

Required. The type of change pending for the ActiveMQ user.

newUserPendingChanges

Arguments

:: ChangeType
$sel:pendingChange:UserPendingChanges'

-> UserPendingChanges

Create a value of UserPendingChanges with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:groups:UserPendingChanges', userPendingChanges_groups - The list of groups (20 maximum) to which the ActiveMQ user belongs. This value can contain only alphanumeric characters, dashes, periods, underscores, and tildes (- . _ ~). This value must be 2-100 characters long.

$sel:consoleAccess:UserPendingChanges', userPendingChanges_consoleAccess - Enables access to the the ActiveMQ Web Console for the ActiveMQ user.

$sel:pendingChange:UserPendingChanges', userPendingChanges_pendingChange - Required. The type of change pending for the ActiveMQ user.

userPendingChanges_groups :: Lens' UserPendingChanges (Maybe [Text])

The list of groups (20 maximum) to which the ActiveMQ user belongs. This value can contain only alphanumeric characters, dashes, periods, underscores, and tildes (- . _ ~). This value must be 2-100 characters long.

userPendingChanges_consoleAccess :: Lens' UserPendingChanges (Maybe Bool)

Enables access to the the ActiveMQ Web Console for the ActiveMQ user.

userPendingChanges_pendingChange :: Lens' UserPendingChanges ChangeType

Required. The type of change pending for the ActiveMQ user.

UserSummary

data UserSummary

Returns a list of all broker users. Does not apply to RabbitMQ brokers.

See: newUserSummary smart constructor.

Constructors

UserSummary'

Fields

pendingChange :: Maybe ChangeType

The type of change pending for the broker user.

username :: Text

Required. The username of the broker user. This value can contain only alphanumeric characters, dashes, periods, underscores, and tildes (- . _ ~). This value must be 2-100 characters long.

newUserSummary

Arguments

:: Text
$sel:username:UserSummary'

-> UserSummary

Create a value of UserSummary with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:pendingChange:UserSummary', userSummary_pendingChange - The type of change pending for the broker user.

$sel:username:UserSummary', userSummary_username - Required. The username of the broker user. This value can contain only alphanumeric characters, dashes, periods, underscores, and tildes (- . _ ~). This value must be 2-100 characters long.

userSummary_pendingChange :: Lens' UserSummary (Maybe ChangeType)

The type of change pending for the broker user.

userSummary_username :: Lens' UserSummary Text

Required. The username of the broker user. This value can contain only alphanumeric characters, dashes, periods, underscores, and tildes (- . _ ~). This value must be 2-100 characters long.

WeeklyStartTime

data WeeklyStartTime

The scheduled time period relative to UTC during which Amazon MQ begins to apply pending updates or patches to the broker.

See: newWeeklyStartTime smart constructor.

Constructors

WeeklyStartTime'

Fields

timeZone :: Maybe Text

The time zone, UTC by default, in either the Country/City format, or the UTC offset format.

timeOfDay :: Text

Required. The time, in 24-hour format.

dayOfWeek :: DayOfWeek

Required. The day of the week.

newWeeklyStartTime

Arguments

:: Text
$sel:timeOfDay:WeeklyStartTime'

-> DayOfWeek
$sel:dayOfWeek:WeeklyStartTime'

-> WeeklyStartTime

Create a value of WeeklyStartTime with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:timeZone:WeeklyStartTime', weeklyStartTime_timeZone - The time zone, UTC by default, in either the Country/City format, or the UTC offset format.

$sel:timeOfDay:WeeklyStartTime', weeklyStartTime_timeOfDay - Required. The time, in 24-hour format.

$sel:dayOfWeek:WeeklyStartTime', weeklyStartTime_dayOfWeek - Required. The day of the week.

weeklyStartTime_timeZone :: Lens' WeeklyStartTime (Maybe Text)

The time zone, UTC by default, in either the Country/City format, or the UTC offset format.

weeklyStartTime_timeOfDay :: Lens' WeeklyStartTime Text

Required. The time, in 24-hour format.

weeklyStartTime_dayOfWeek :: Lens' WeeklyStartTime DayOfWeek

Required. The day of the week.