Amazonka.MQ.Types.BrokerSummary

Description

Documentation

data BrokerSummary

Returns information about all brokers.

See: newBrokerSummary smart constructor.

Constructors

BrokerSummary'

Fields

brokerName :: Maybe Text

The broker's name. This value is unique in your AWS account, 1-50 characters long, and containing only letters, numbers, dashes, and underscores, and must not contain white spaces, brackets, wildcard characters, or special characters.

brokerState :: Maybe BrokerState

The broker's status.

created :: Maybe POSIX

The time when the broker was created.

brokerId :: Maybe Text

The unique ID that Amazon MQ generates for the broker.

brokerArn :: Maybe Text

The broker's Amazon Resource Name (ARN).

hostInstanceType :: Maybe Text

The broker's instance type.

deploymentMode :: DeploymentMode

The broker's deployment mode.

engineType :: EngineType

The type of broker engine.

newBrokerSummary

Arguments

:: DeploymentMode
$sel:deploymentMode:BrokerSummary'

-> EngineType
$sel:engineType:BrokerSummary'

-> BrokerSummary

Create a value of BrokerSummary with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:brokerName:BrokerSummary', brokerSummary_brokerName - The broker's name. This value is unique in your AWS account, 1-50 characters long, and containing only letters, numbers, dashes, and underscores, and must not contain white spaces, brackets, wildcard characters, or special characters.

$sel:brokerState:BrokerSummary', brokerSummary_brokerState - The broker's status.

$sel:created:BrokerSummary', brokerSummary_created - The time when the broker was created.

$sel:brokerId:BrokerSummary', brokerSummary_brokerId - The unique ID that Amazon MQ generates for the broker.

$sel:brokerArn:BrokerSummary', brokerSummary_brokerArn - The broker's Amazon Resource Name (ARN).

$sel:hostInstanceType:BrokerSummary', brokerSummary_hostInstanceType - The broker's instance type.

$sel:deploymentMode:BrokerSummary', brokerSummary_deploymentMode - The broker's deployment mode.

$sel:engineType:BrokerSummary', brokerSummary_engineType - The type of broker engine.

brokerSummary_brokerName :: Lens' BrokerSummary (Maybe Text)

The broker's name. This value is unique in your AWS account, 1-50 characters long, and containing only letters, numbers, dashes, and underscores, and must not contain white spaces, brackets, wildcard characters, or special characters.

brokerSummary_brokerState :: Lens' BrokerSummary (Maybe BrokerState)

The broker's status.

brokerSummary_created :: Lens' BrokerSummary (Maybe UTCTime)

The time when the broker was created.

brokerSummary_brokerId :: Lens' BrokerSummary (Maybe Text)

The unique ID that Amazon MQ generates for the broker.

brokerSummary_brokerArn :: Lens' BrokerSummary (Maybe Text)

The broker's Amazon Resource Name (ARN).

brokerSummary_hostInstanceType :: Lens' BrokerSummary (Maybe Text)

The broker's instance type.

brokerSummary_deploymentMode :: Lens' BrokerSummary DeploymentMode

The broker's deployment mode.

brokerSummary_engineType :: Lens' BrokerSummary EngineType

The type of broker engine.