Amazonka.MQ.Types.Logs

Description

Documentation

data Logs

The list of information about logs to be enabled for the specified broker.

See: newLogs smart constructor.

Constructors

Logs'

Fields

audit :: Maybe Bool

Enables audit logging. Every user management action made using JMX or the ActiveMQ Web Console is logged. Does not apply to RabbitMQ brokers.

general :: Maybe Bool

Enables general logging.

newLogs :: Logs

Create a value of Logs with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:audit:Logs', logs_audit - Enables audit logging. Every user management action made using JMX or the ActiveMQ Web Console is logged. Does not apply to RabbitMQ brokers.

$sel:general:Logs', logs_general - Enables general logging.

logs_audit :: Lens' Logs (Maybe Bool)

Enables audit logging. Every user management action made using JMX or the ActiveMQ Web Console is logged. Does not apply to RabbitMQ brokers.

logs_general :: Lens' Logs (Maybe Bool)

Enables general logging.