Amazonka.MQ.UpdateConfiguration

Description

Updates the specified configuration.

Creating a Request

data UpdateConfiguration

Updates the specified configuration.

See: newUpdateConfiguration smart constructor.

Constructors

UpdateConfiguration'

Fields

description :: Maybe Text

The description of the configuration.

configurationId :: Text

The unique ID that Amazon MQ generates for the configuration.

data' :: Text

Required. The base64-encoded XML configuration.

newUpdateConfiguration

Arguments

:: Text
$sel:configurationId:UpdateConfiguration'

-> Text
$sel:data':UpdateConfiguration'

-> UpdateConfiguration

Create a value of UpdateConfiguration with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:description:UpdateConfiguration', updateConfiguration_description - The description of the configuration.

$sel:configurationId:UpdateConfiguration', updateConfiguration_configurationId - The unique ID that Amazon MQ generates for the configuration.

$sel:data':UpdateConfiguration', updateConfiguration_data - Required. The base64-encoded XML configuration.

Request Lenses

updateConfiguration_description :: Lens' UpdateConfiguration (Maybe Text)

The description of the configuration.

updateConfiguration_configurationId :: Lens' UpdateConfiguration Text

The unique ID that Amazon MQ generates for the configuration.

updateConfiguration_data :: Lens' UpdateConfiguration Text

Required. The base64-encoded XML configuration.

Destructuring the Response

data UpdateConfigurationResponse

See: newUpdateConfigurationResponse smart constructor.

Constructors

UpdateConfigurationResponse'

Fields

arn :: Maybe Text

Required. The Amazon Resource Name (ARN) of the configuration.

latestRevision :: Maybe ConfigurationRevision

The latest revision of the configuration.

created :: Maybe POSIX

Required. The date and time of the configuration.

warnings :: Maybe [SanitizationWarning]

The list of the first 20 warnings about the configuration XML elements or attributes that were sanitized.

name :: Maybe Text

Required. The name of the configuration. This value can contain only alphanumeric characters, dashes, periods, underscores, and tildes (- . _ ~). This value must be 1-150 characters long.

id :: Maybe Text

Required. The unique ID that Amazon MQ generates for the configuration.

httpStatus :: Int

The response's http status code.

newUpdateConfigurationResponse

Arguments

:: Int
$sel:httpStatus:UpdateConfigurationResponse'

-> UpdateConfigurationResponse

Create a value of UpdateConfigurationResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:arn:UpdateConfigurationResponse', updateConfigurationResponse_arn - Required. The Amazon Resource Name (ARN) of the configuration.

$sel:latestRevision:UpdateConfigurationResponse', updateConfigurationResponse_latestRevision - The latest revision of the configuration.

$sel:created:UpdateConfigurationResponse', updateConfigurationResponse_created - Required. The date and time of the configuration.

$sel:warnings:UpdateConfigurationResponse', updateConfigurationResponse_warnings - The list of the first 20 warnings about the configuration XML elements or attributes that were sanitized.

$sel:name:UpdateConfigurationResponse', updateConfigurationResponse_name - Required. The name of the configuration. This value can contain only alphanumeric characters, dashes, periods, underscores, and tildes (- . _ ~). This value must be 1-150 characters long.

$sel:id:UpdateConfigurationResponse', updateConfigurationResponse_id - Required. The unique ID that Amazon MQ generates for the configuration.

$sel:httpStatus:UpdateConfigurationResponse', updateConfigurationResponse_httpStatus - The response's http status code.

Response Lenses

updateConfigurationResponse_arn :: Lens' UpdateConfigurationResponse (Maybe Text)

Required. The Amazon Resource Name (ARN) of the configuration.

updateConfigurationResponse_latestRevision :: Lens' UpdateConfigurationResponse (Maybe ConfigurationRevision)

The latest revision of the configuration.

updateConfigurationResponse_created :: Lens' UpdateConfigurationResponse (Maybe UTCTime)

Required. The date and time of the configuration.

updateConfigurationResponse_warnings :: Lens' UpdateConfigurationResponse (Maybe [SanitizationWarning])

The list of the first 20 warnings about the configuration XML elements or attributes that were sanitized.

updateConfigurationResponse_name :: Lens' UpdateConfigurationResponse (Maybe Text)

Required. The name of the configuration. This value can contain only alphanumeric characters, dashes, periods, underscores, and tildes (- . _ ~). This value must be 1-150 characters long.

updateConfigurationResponse_id :: Lens' UpdateConfigurationResponse (Maybe Text)

Required. The unique ID that Amazon MQ generates for the configuration.

updateConfigurationResponse_httpStatus :: Lens' UpdateConfigurationResponse Int

The response's http status code.