Amazonka.MachineLearning.AddTags

Description

Adds one or more tags to an object, up to a limit of 10. Each tag consists of a key and an optional value. If you add a tag using a key that is already associated with the ML object, AddTags updates the tag's value.

Creating a Request

data AddTags

See: newAddTags smart constructor.

Constructors

AddTags'

Fields

tags :: [Tag]

The key-value pairs to use to create tags. If you specify a key without specifying a value, Amazon ML creates a tag with the specified key and a value of null.

resourceId :: Text

The ID of the ML object to tag. For example, exampleModelId.

resourceType :: TaggableResourceType

The type of the ML object to tag.

newAddTags

Arguments

:: Text
$sel:resourceId:AddTags'

-> TaggableResourceType
$sel:resourceType:AddTags'

-> AddTags

Create a value of AddTags with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:tags:AddTags', addTags_tags - The key-value pairs to use to create tags. If you specify a key without specifying a value, Amazon ML creates a tag with the specified key and a value of null.

$sel:resourceId:AddTags', addTags_resourceId - The ID of the ML object to tag. For example, exampleModelId.

$sel:resourceType:AddTags', addTags_resourceType - The type of the ML object to tag.

Request Lenses

addTags_tags :: Lens' AddTags [Tag]

The key-value pairs to use to create tags. If you specify a key without specifying a value, Amazon ML creates a tag with the specified key and a value of null.

addTags_resourceId :: Lens' AddTags Text

The ID of the ML object to tag. For example, exampleModelId.

addTags_resourceType :: Lens' AddTags TaggableResourceType

The type of the ML object to tag.

Destructuring the Response

data AddTagsResponse

Amazon ML returns the following elements.

See: newAddTagsResponse smart constructor.

Constructors

AddTagsResponse'

Fields

resourceId :: Maybe Text

The ID of the ML object that was tagged.

resourceType :: Maybe TaggableResourceType

The type of the ML object that was tagged.

httpStatus :: Int

The response's http status code.

newAddTagsResponse

Arguments

:: Int
$sel:httpStatus:AddTagsResponse'

-> AddTagsResponse

Create a value of AddTagsResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:resourceId:AddTags', addTagsResponse_resourceId - The ID of the ML object that was tagged.

$sel:resourceType:AddTags', addTagsResponse_resourceType - The type of the ML object that was tagged.

$sel:httpStatus:AddTagsResponse', addTagsResponse_httpStatus - The response's http status code.

Response Lenses

addTagsResponse_resourceId :: Lens' AddTagsResponse (Maybe Text)

The ID of the ML object that was tagged.

addTagsResponse_resourceType :: Lens' AddTagsResponse (Maybe TaggableResourceType)

The type of the ML object that was tagged.

addTagsResponse_httpStatus :: Lens' AddTagsResponse Int

The response's http status code.