Amazonka.MachineLearning.CreateBatchPrediction

Description

Generates predictions for a group of observations. The observations to process exist in one or more data files referenced by a DataSource. This operation creates a new BatchPrediction, and uses an MLModel and the data files referenced by the DataSource as information sources.

CreateBatchPrediction is an asynchronous operation. In response to CreateBatchPrediction, Amazon Machine Learning (Amazon ML) immediately returns and sets the BatchPrediction status to PENDING. After the BatchPrediction completes, Amazon ML sets the status to COMPLETED.

You can poll for status updates by using the GetBatchPrediction operation and checking the Status parameter of the result. After the COMPLETED status appears, the results are available in the location specified by the OutputUri parameter.

Creating a Request

data CreateBatchPrediction

See: newCreateBatchPrediction smart constructor.

Constructors

CreateBatchPrediction'

Fields

batchPredictionName :: Maybe Text

A user-supplied name or description of the BatchPrediction. BatchPredictionName can only use the UTF-8 character set.

batchPredictionId :: Text

A user-supplied ID that uniquely identifies the BatchPrediction.

mLModelId :: Text

The ID of the MLModel that will generate predictions for the group of observations.

batchPredictionDataSourceId :: Text

The ID of the DataSource that points to the group of observations to predict.

outputUri :: Text

The location of an Amazon Simple Storage Service (Amazon S3) bucket or directory to store the batch prediction results. The following substrings are not allowed in the s3 key portion of the outputURI field: ':', '//', '/./', '/../'.

Amazon ML needs permissions to store and retrieve the logs on your behalf. For information about how to set permissions, see the Amazon Machine Learning Developer Guide.

newCreateBatchPrediction

Arguments

:: Text
$sel:batchPredictionId:CreateBatchPrediction'

-> Text
$sel:mLModelId:CreateBatchPrediction'

-> Text
$sel:batchPredictionDataSourceId:CreateBatchPrediction'

-> Text
$sel:outputUri:CreateBatchPrediction'

-> CreateBatchPrediction

Create a value of CreateBatchPrediction with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:batchPredictionName:CreateBatchPrediction', createBatchPrediction_batchPredictionName - A user-supplied name or description of the BatchPrediction. BatchPredictionName can only use the UTF-8 character set.

$sel:batchPredictionId:CreateBatchPrediction', createBatchPrediction_batchPredictionId - A user-supplied ID that uniquely identifies the BatchPrediction.

$sel:mLModelId:CreateBatchPrediction', createBatchPrediction_mLModelId - The ID of the MLModel that will generate predictions for the group of observations.

$sel:batchPredictionDataSourceId:CreateBatchPrediction', createBatchPrediction_batchPredictionDataSourceId - The ID of the DataSource that points to the group of observations to predict.

$sel:outputUri:CreateBatchPrediction', createBatchPrediction_outputUri - The location of an Amazon Simple Storage Service (Amazon S3) bucket or directory to store the batch prediction results. The following substrings are not allowed in the s3 key portion of the outputURI field: ':', '//', '/./', '/../'.

Amazon ML needs permissions to store and retrieve the logs on your behalf. For information about how to set permissions, see the Amazon Machine Learning Developer Guide.

Request Lenses

createBatchPrediction_batchPredictionName :: Lens' CreateBatchPrediction (Maybe Text)

A user-supplied name or description of the BatchPrediction. BatchPredictionName can only use the UTF-8 character set.

createBatchPrediction_batchPredictionId :: Lens' CreateBatchPrediction Text

A user-supplied ID that uniquely identifies the BatchPrediction.

createBatchPrediction_mLModelId :: Lens' CreateBatchPrediction Text

The ID of the MLModel that will generate predictions for the group of observations.

createBatchPrediction_batchPredictionDataSourceId :: Lens' CreateBatchPrediction Text

The ID of the DataSource that points to the group of observations to predict.

createBatchPrediction_outputUri :: Lens' CreateBatchPrediction Text

The location of an Amazon Simple Storage Service (Amazon S3) bucket or directory to store the batch prediction results. The following substrings are not allowed in the s3 key portion of the outputURI field: ':', '//', '/./', '/../'.

Amazon ML needs permissions to store and retrieve the logs on your behalf. For information about how to set permissions, see the Amazon Machine Learning Developer Guide.

Destructuring the Response

data CreateBatchPredictionResponse

Represents the output of a CreateBatchPrediction operation, and is an acknowledgement that Amazon ML received the request.

The CreateBatchPrediction operation is asynchronous. You can poll for status updates by using the >GetBatchPrediction operation and checking the Status parameter of the result.

See: newCreateBatchPredictionResponse smart constructor.

Constructors

CreateBatchPredictionResponse'

Fields

batchPredictionId :: Maybe Text

A user-supplied ID that uniquely identifies the BatchPrediction. This value is identical to the value of the BatchPredictionId in the request.

httpStatus :: Int

The response's http status code.

newCreateBatchPredictionResponse

Arguments

:: Int
$sel:httpStatus:CreateBatchPredictionResponse'

-> CreateBatchPredictionResponse

Create a value of CreateBatchPredictionResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:batchPredictionId:CreateBatchPrediction', createBatchPredictionResponse_batchPredictionId - A user-supplied ID that uniquely identifies the BatchPrediction. This value is identical to the value of the BatchPredictionId in the request.

$sel:httpStatus:CreateBatchPredictionResponse', createBatchPredictionResponse_httpStatus - The response's http status code.

Response Lenses

createBatchPredictionResponse_batchPredictionId :: Lens' CreateBatchPredictionResponse (Maybe Text)

A user-supplied ID that uniquely identifies the BatchPrediction. This value is identical to the value of the BatchPredictionId in the request.

createBatchPredictionResponse_httpStatus :: Lens' CreateBatchPredictionResponse Int

The response's http status code.