Amazonka.MachineLearning.DescribeDataSources

Description

Returns a list of DataSource that match the search criteria in the request.

This operation returns paginated results.

Creating a Request

data DescribeDataSources

See: newDescribeDataSources smart constructor.

Constructors

DescribeDataSources'

Fields

eq :: Maybe Text

The equal to operator. The DataSource results will have FilterVariable values that exactly match the value specified with EQ.

ge :: Maybe Text

The greater than or equal to operator. The DataSource results will have FilterVariable values that are greater than or equal to the value specified with GE.

prefix :: Maybe Text

A string that is found at the beginning of a variable, such as Name or Id.

For example, a DataSource could have the Name 2014-09-09-HolidayGiftMailer. To search for this DataSource, select Name for the FilterVariable and any of the following strings for the Prefix:

2014-09

2014-09-09

2014-09-09-Holiday

gt :: Maybe Text

The greater than operator. The DataSource results will have FilterVariable values that are greater than the value specified with GT.

ne :: Maybe Text

The not equal to operator. The DataSource results will have FilterVariable values not equal to the value specified with NE.

nextToken :: Maybe Text

The ID of the page in the paginated results.

sortOrder :: Maybe SortOrder

A two-value parameter that determines the sequence of the resulting list of DataSource.

asc - Arranges the list in ascending order (A-Z, 0-9).

dsc - Arranges the list in descending order (Z-A, 9-0).

Results are sorted by FilterVariable.

limit :: Maybe Natural

The maximum number of DataSource to include in the result.

lt :: Maybe Text

The less than operator. The DataSource results will have FilterVariable values that are less than the value specified with LT.

filterVariable :: Maybe DataSourceFilterVariable

Use one of the following variables to filter a list of DataSource:

CreatedAt - Sets the search criteria to DataSource creation dates.

Status - Sets the search criteria to DataSource statuses.

Name - Sets the search criteria to the contents of DataSource Name.

DataUri - Sets the search criteria to the URI of data files used to create the DataSource. The URI can identify either a file or an Amazon Simple Storage Service (Amazon S3) bucket or directory.

IAMUser - Sets the search criteria to the user account that invoked the DataSource creation.

le :: Maybe Text

The less than or equal to operator. The DataSource results will have FilterVariable values that are less than or equal to the value specified with LE.

newDescribeDataSources :: DescribeDataSources

Create a value of DescribeDataSources with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:eq:DescribeDataSources', describeDataSources_eq - The equal to operator. The DataSource results will have FilterVariable values that exactly match the value specified with EQ.

$sel:ge:DescribeDataSources', describeDataSources_ge - The greater than or equal to operator. The DataSource results will have FilterVariable values that are greater than or equal to the value specified with GE.

$sel:prefix:DescribeDataSources', describeDataSources_prefix - A string that is found at the beginning of a variable, such as Name or Id.

For example, a DataSource could have the Name 2014-09-09-HolidayGiftMailer. To search for this DataSource, select Name for the FilterVariable and any of the following strings for the Prefix:

2014-09

2014-09-09

2014-09-09-Holiday

$sel:gt:DescribeDataSources', describeDataSources_gt - The greater than operator. The DataSource results will have FilterVariable values that are greater than the value specified with GT.

$sel:ne:DescribeDataSources', describeDataSources_ne - The not equal to operator. The DataSource results will have FilterVariable values not equal to the value specified with NE.

$sel:nextToken:DescribeDataSources', describeDataSources_nextToken - The ID of the page in the paginated results.

$sel:sortOrder:DescribeDataSources', describeDataSources_sortOrder - A two-value parameter that determines the sequence of the resulting list of DataSource.

asc - Arranges the list in ascending order (A-Z, 0-9).

dsc - Arranges the list in descending order (Z-A, 9-0).

Results are sorted by FilterVariable.

$sel:limit:DescribeDataSources', describeDataSources_limit - The maximum number of DataSource to include in the result.

$sel:lt:DescribeDataSources', describeDataSources_lt - The less than operator. The DataSource results will have FilterVariable values that are less than the value specified with LT.

$sel:filterVariable:DescribeDataSources', describeDataSources_filterVariable - Use one of the following variables to filter a list of DataSource:

CreatedAt - Sets the search criteria to DataSource creation dates.

Status - Sets the search criteria to DataSource statuses.

Name - Sets the search criteria to the contents of DataSource Name.

DataUri - Sets the search criteria to the URI of data files used to create the DataSource. The URI can identify either a file or an Amazon Simple Storage Service (Amazon S3) bucket or directory.

IAMUser - Sets the search criteria to the user account that invoked the DataSource creation.

$sel:le:DescribeDataSources', describeDataSources_le - The less than or equal to operator. The DataSource results will have FilterVariable values that are less than or equal to the value specified with LE.

Request Lenses

describeDataSources_eq :: Lens' DescribeDataSources (Maybe Text)

The equal to operator. The DataSource results will have FilterVariable values that exactly match the value specified with EQ.

describeDataSources_ge :: Lens' DescribeDataSources (Maybe Text)

The greater than or equal to operator. The DataSource results will have FilterVariable values that are greater than or equal to the value specified with GE.

describeDataSources_prefix :: Lens' DescribeDataSources (Maybe Text)

A string that is found at the beginning of a variable, such as Name or Id.

For example, a DataSource could have the Name 2014-09-09-HolidayGiftMailer. To search for this DataSource, select Name for the FilterVariable and any of the following strings for the Prefix:

2014-09

2014-09-09

2014-09-09-Holiday

describeDataSources_gt :: Lens' DescribeDataSources (Maybe Text)

The greater than operator. The DataSource results will have FilterVariable values that are greater than the value specified with GT.

describeDataSources_ne :: Lens' DescribeDataSources (Maybe Text)

The not equal to operator. The DataSource results will have FilterVariable values not equal to the value specified with NE.

describeDataSources_nextToken :: Lens' DescribeDataSources (Maybe Text)

The ID of the page in the paginated results.

describeDataSources_sortOrder :: Lens' DescribeDataSources (Maybe SortOrder)

A two-value parameter that determines the sequence of the resulting list of DataSource.

asc - Arranges the list in ascending order (A-Z, 0-9).

dsc - Arranges the list in descending order (Z-A, 9-0).

Results are sorted by FilterVariable.

describeDataSources_limit :: Lens' DescribeDataSources (Maybe Natural)

The maximum number of DataSource to include in the result.

describeDataSources_lt :: Lens' DescribeDataSources (Maybe Text)

The less than operator. The DataSource results will have FilterVariable values that are less than the value specified with LT.

describeDataSources_filterVariable :: Lens' DescribeDataSources (Maybe DataSourceFilterVariable)

Use one of the following variables to filter a list of DataSource:

CreatedAt - Sets the search criteria to DataSource creation dates.

Status - Sets the search criteria to DataSource statuses.

Name - Sets the search criteria to the contents of DataSource Name.

DataUri - Sets the search criteria to the URI of data files used to create the DataSource. The URI can identify either a file or an Amazon Simple Storage Service (Amazon S3) bucket or directory.

IAMUser - Sets the search criteria to the user account that invoked the DataSource creation.

describeDataSources_le :: Lens' DescribeDataSources (Maybe Text)

The less than or equal to operator. The DataSource results will have FilterVariable values that are less than or equal to the value specified with LE.

Destructuring the Response

data DescribeDataSourcesResponse

Represents the query results from a DescribeDataSources operation. The content is essentially a list of DataSource.

See: newDescribeDataSourcesResponse smart constructor.

Constructors

DescribeDataSourcesResponse'

Fields

results :: Maybe [DataSource]

A list of DataSource that meet the search criteria.

nextToken :: Maybe Text

An ID of the next page in the paginated results that indicates at least one more page follows.

httpStatus :: Int

The response's http status code.

newDescribeDataSourcesResponse

Arguments

:: Int
$sel:httpStatus:DescribeDataSourcesResponse'

-> DescribeDataSourcesResponse

Create a value of DescribeDataSourcesResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:results:DescribeDataSourcesResponse', describeDataSourcesResponse_results - A list of DataSource that meet the search criteria.

$sel:nextToken:DescribeDataSources', describeDataSourcesResponse_nextToken - An ID of the next page in the paginated results that indicates at least one more page follows.

$sel:httpStatus:DescribeDataSourcesResponse', describeDataSourcesResponse_httpStatus - The response's http status code.

Response Lenses

describeDataSourcesResponse_results :: Lens' DescribeDataSourcesResponse (Maybe [DataSource])

A list of DataSource that meet the search criteria.

describeDataSourcesResponse_nextToken :: Lens' DescribeDataSourcesResponse (Maybe Text)

An ID of the next page in the paginated results that indicates at least one more page follows.

describeDataSourcesResponse_httpStatus :: Lens' DescribeDataSourcesResponse Int

The response's http status code.