Amazonka.MachineLearning.GetBatchPrediction

Description

Returns a BatchPrediction that includes detailed metadata, status, and data file information for a Batch Prediction request.

Creating a Request

data GetBatchPrediction

See: newGetBatchPrediction smart constructor.

Constructors

GetBatchPrediction'

Fields

batchPredictionId :: Text

An ID assigned to the BatchPrediction at creation.

newGetBatchPrediction

Arguments

:: Text
$sel:batchPredictionId:GetBatchPrediction'

-> GetBatchPrediction

Create a value of GetBatchPrediction with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:batchPredictionId:GetBatchPrediction', getBatchPrediction_batchPredictionId - An ID assigned to the BatchPrediction at creation.

Request Lenses

getBatchPrediction_batchPredictionId :: Lens' GetBatchPrediction Text

An ID assigned to the BatchPrediction at creation.

Destructuring the Response

data GetBatchPredictionResponse

Represents the output of a GetBatchPrediction operation and describes a BatchPrediction.

See: newGetBatchPredictionResponse smart constructor.

Constructors

GetBatchPredictionResponse'

Fields

status :: Maybe EntityStatus

The status of the BatchPrediction, which can be one of the following values:

PENDING - Amazon Machine Learning (Amazon ML) submitted a request to generate batch predictions.

INPROGRESS - The batch predictions are in progress.

FAILED - The request to perform a batch prediction did not run to completion. It is not usable.

COMPLETED - The batch prediction process completed successfully.

DELETED - The BatchPrediction is marked as deleted. It is not usable.

lastUpdatedAt :: Maybe POSIX

The time of the most recent edit to BatchPrediction. The time is expressed in epoch time.

createdAt :: Maybe POSIX

The time when the BatchPrediction was created. The time is expressed in epoch time.

computeTime :: Maybe Integer

The approximate CPU time in milliseconds that Amazon Machine Learning spent processing the BatchPrediction, normalized and scaled on computation resources. ComputeTime is only available if the BatchPrediction is in the COMPLETED state.

inputDataLocationS3 :: Maybe Text

The location of the data file or directory in Amazon Simple Storage Service (Amazon S3).

mLModelId :: Maybe Text

The ID of the MLModel that generated predictions for the BatchPrediction request.

batchPredictionDataSourceId :: Maybe Text

The ID of the DataSource that was used to create the BatchPrediction.

totalRecordCount :: Maybe Integer

The number of total records that Amazon Machine Learning saw while processing the BatchPrediction.

startedAt :: Maybe POSIX

The epoch time when Amazon Machine Learning marked the BatchPrediction as INPROGRESS. StartedAt isn't available if the BatchPrediction is in the PENDING state.

batchPredictionId :: Maybe Text

An ID assigned to the BatchPrediction at creation. This value should be identical to the value of the BatchPredictionID in the request.

finishedAt :: Maybe POSIX

The epoch time when Amazon Machine Learning marked the BatchPrediction as COMPLETED or FAILED. FinishedAt is only available when the BatchPrediction is in the COMPLETED or FAILED state.

invalidRecordCount :: Maybe Integer

The number of invalid records that Amazon Machine Learning saw while processing the BatchPrediction.

createdByIamUser :: Maybe Text

The AWS user account that invoked the BatchPrediction. The account type can be either an AWS root account or an AWS Identity and Access Management (IAM) user account.

name :: Maybe Text

A user-supplied name or description of the BatchPrediction.

logUri :: Maybe Text

A link to the file that contains logs of the CreateBatchPrediction operation.

message :: Maybe Text

A description of the most recent details about processing the batch prediction request.

outputUri :: Maybe Text

The location of an Amazon S3 bucket or directory to receive the operation results.

httpStatus :: Int

The response's http status code.

newGetBatchPredictionResponse

Arguments

:: Int
$sel:httpStatus:GetBatchPredictionResponse'

-> GetBatchPredictionResponse

Create a value of GetBatchPredictionResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:status:GetBatchPredictionResponse', getBatchPredictionResponse_status - The status of the BatchPrediction, which can be one of the following values:

PENDING - Amazon Machine Learning (Amazon ML) submitted a request to generate batch predictions.

INPROGRESS - The batch predictions are in progress.

FAILED - The request to perform a batch prediction did not run to completion. It is not usable.

COMPLETED - The batch prediction process completed successfully.

DELETED - The BatchPrediction is marked as deleted. It is not usable.

$sel:lastUpdatedAt:GetBatchPredictionResponse', getBatchPredictionResponse_lastUpdatedAt - The time of the most recent edit to BatchPrediction. The time is expressed in epoch time.

$sel:createdAt:GetBatchPredictionResponse', getBatchPredictionResponse_createdAt - The time when the BatchPrediction was created. The time is expressed in epoch time.

$sel:computeTime:GetBatchPredictionResponse', getBatchPredictionResponse_computeTime - The approximate CPU time in milliseconds that Amazon Machine Learning spent processing the BatchPrediction, normalized and scaled on computation resources. ComputeTime is only available if the BatchPrediction is in the COMPLETED state.

$sel:inputDataLocationS3:GetBatchPredictionResponse', getBatchPredictionResponse_inputDataLocationS3 - The location of the data file or directory in Amazon Simple Storage Service (Amazon S3).

$sel:mLModelId:GetBatchPredictionResponse', getBatchPredictionResponse_mLModelId - The ID of the MLModel that generated predictions for the BatchPrediction request.

$sel:batchPredictionDataSourceId:GetBatchPredictionResponse', getBatchPredictionResponse_batchPredictionDataSourceId - The ID of the DataSource that was used to create the BatchPrediction.

$sel:totalRecordCount:GetBatchPredictionResponse', getBatchPredictionResponse_totalRecordCount - The number of total records that Amazon Machine Learning saw while processing the BatchPrediction.

$sel:startedAt:GetBatchPredictionResponse', getBatchPredictionResponse_startedAt - The epoch time when Amazon Machine Learning marked the BatchPrediction as INPROGRESS. StartedAt isn't available if the BatchPrediction is in the PENDING state.

$sel:batchPredictionId:GetBatchPrediction', getBatchPredictionResponse_batchPredictionId - An ID assigned to the BatchPrediction at creation. This value should be identical to the value of the BatchPredictionID in the request.

$sel:finishedAt:GetBatchPredictionResponse', getBatchPredictionResponse_finishedAt - The epoch time when Amazon Machine Learning marked the BatchPrediction as COMPLETED or FAILED. FinishedAt is only available when the BatchPrediction is in the COMPLETED or FAILED state.

$sel:invalidRecordCount:GetBatchPredictionResponse', getBatchPredictionResponse_invalidRecordCount - The number of invalid records that Amazon Machine Learning saw while processing the BatchPrediction.

$sel:createdByIamUser:GetBatchPredictionResponse', getBatchPredictionResponse_createdByIamUser - The AWS user account that invoked the BatchPrediction. The account type can be either an AWS root account or an AWS Identity and Access Management (IAM) user account.

$sel:name:GetBatchPredictionResponse', getBatchPredictionResponse_name - A user-supplied name or description of the BatchPrediction.

$sel:logUri:GetBatchPredictionResponse', getBatchPredictionResponse_logUri - A link to the file that contains logs of the CreateBatchPrediction operation.

$sel:message:GetBatchPredictionResponse', getBatchPredictionResponse_message - A description of the most recent details about processing the batch prediction request.

$sel:outputUri:GetBatchPredictionResponse', getBatchPredictionResponse_outputUri - The location of an Amazon S3 bucket or directory to receive the operation results.

$sel:httpStatus:GetBatchPredictionResponse', getBatchPredictionResponse_httpStatus - The response's http status code.

Response Lenses

getBatchPredictionResponse_status :: Lens' GetBatchPredictionResponse (Maybe EntityStatus)

The status of the BatchPrediction, which can be one of the following values:

PENDING - Amazon Machine Learning (Amazon ML) submitted a request to generate batch predictions.

INPROGRESS - The batch predictions are in progress.

FAILED - The request to perform a batch prediction did not run to completion. It is not usable.

COMPLETED - The batch prediction process completed successfully.

DELETED - The BatchPrediction is marked as deleted. It is not usable.

getBatchPredictionResponse_lastUpdatedAt :: Lens' GetBatchPredictionResponse (Maybe UTCTime)

The time of the most recent edit to BatchPrediction. The time is expressed in epoch time.

getBatchPredictionResponse_createdAt :: Lens' GetBatchPredictionResponse (Maybe UTCTime)

The time when the BatchPrediction was created. The time is expressed in epoch time.

getBatchPredictionResponse_computeTime :: Lens' GetBatchPredictionResponse (Maybe Integer)

The approximate CPU time in milliseconds that Amazon Machine Learning spent processing the BatchPrediction, normalized and scaled on computation resources. ComputeTime is only available if the BatchPrediction is in the COMPLETED state.

getBatchPredictionResponse_inputDataLocationS3 :: Lens' GetBatchPredictionResponse (Maybe Text)

The location of the data file or directory in Amazon Simple Storage Service (Amazon S3).

getBatchPredictionResponse_mLModelId :: Lens' GetBatchPredictionResponse (Maybe Text)

The ID of the MLModel that generated predictions for the BatchPrediction request.

getBatchPredictionResponse_batchPredictionDataSourceId :: Lens' GetBatchPredictionResponse (Maybe Text)

The ID of the DataSource that was used to create the BatchPrediction.

getBatchPredictionResponse_totalRecordCount :: Lens' GetBatchPredictionResponse (Maybe Integer)

The number of total records that Amazon Machine Learning saw while processing the BatchPrediction.

getBatchPredictionResponse_startedAt :: Lens' GetBatchPredictionResponse (Maybe UTCTime)

The epoch time when Amazon Machine Learning marked the BatchPrediction as INPROGRESS. StartedAt isn't available if the BatchPrediction is in the PENDING state.

getBatchPredictionResponse_batchPredictionId :: Lens' GetBatchPredictionResponse (Maybe Text)

An ID assigned to the BatchPrediction at creation. This value should be identical to the value of the BatchPredictionID in the request.

getBatchPredictionResponse_finishedAt :: Lens' GetBatchPredictionResponse (Maybe UTCTime)

The epoch time when Amazon Machine Learning marked the BatchPrediction as COMPLETED or FAILED. FinishedAt is only available when the BatchPrediction is in the COMPLETED or FAILED state.

getBatchPredictionResponse_invalidRecordCount :: Lens' GetBatchPredictionResponse (Maybe Integer)

The number of invalid records that Amazon Machine Learning saw while processing the BatchPrediction.

getBatchPredictionResponse_createdByIamUser :: Lens' GetBatchPredictionResponse (Maybe Text)

The AWS user account that invoked the BatchPrediction. The account type can be either an AWS root account or an AWS Identity and Access Management (IAM) user account.

getBatchPredictionResponse_name :: Lens' GetBatchPredictionResponse (Maybe Text)

A user-supplied name or description of the BatchPrediction.

getBatchPredictionResponse_logUri :: Lens' GetBatchPredictionResponse (Maybe Text)

A link to the file that contains logs of the CreateBatchPrediction operation.

getBatchPredictionResponse_message :: Lens' GetBatchPredictionResponse (Maybe Text)

A description of the most recent details about processing the batch prediction request.

getBatchPredictionResponse_outputUri :: Lens' GetBatchPredictionResponse (Maybe Text)

The location of an Amazon S3 bucket or directory to receive the operation results.

getBatchPredictionResponse_httpStatus :: Lens' GetBatchPredictionResponse Int

The response's http status code.