Amazonka.MachineLearning.GetEvaluation

Description

Returns an Evaluation that includes metadata as well as the current status of the Evaluation.

Creating a Request

data GetEvaluation

See: newGetEvaluation smart constructor.

Constructors

GetEvaluation'

Fields

evaluationId :: Text

The ID of the Evaluation to retrieve. The evaluation of each MLModel is recorded and cataloged. The ID provides the means to access the information.

newGetEvaluation

Arguments

:: Text
$sel:evaluationId:GetEvaluation'

-> GetEvaluation

Create a value of GetEvaluation with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:evaluationId:GetEvaluation', getEvaluation_evaluationId - The ID of the Evaluation to retrieve. The evaluation of each MLModel is recorded and cataloged. The ID provides the means to access the information.

Request Lenses

getEvaluation_evaluationId :: Lens' GetEvaluation Text

The ID of the Evaluation to retrieve. The evaluation of each MLModel is recorded and cataloged. The ID provides the means to access the information.

Destructuring the Response

data GetEvaluationResponse

Represents the output of a GetEvaluation operation and describes an Evaluation.

See: newGetEvaluationResponse smart constructor.

Constructors

GetEvaluationResponse'

Fields

status :: Maybe EntityStatus

The status of the evaluation. This element can have one of the following values:

PENDING - Amazon Machine Language (Amazon ML) submitted a request to evaluate an MLModel.

INPROGRESS - The evaluation is underway.

FAILED - The request to evaluate an MLModel did not run to completion. It is not usable.

COMPLETED - The evaluation process completed successfully.

DELETED - The Evaluation is marked as deleted. It is not usable.

performanceMetrics :: Maybe PerformanceMetrics

Measurements of how well the MLModel performed using observations referenced by the DataSource. One of the following metric is returned based on the type of the MLModel:

BinaryAUC: A binary MLModel uses the Area Under the Curve (AUC) technique to measure performance.

RegressionRMSE: A regression MLModel uses the Root Mean Square Error (RMSE) technique to measure performance. RMSE measures the difference between predicted and actual values for a single variable.

MulticlassAvgFScore: A multiclass MLModel uses the F1 score technique to measure performance.

For more information about performance metrics, please see the Amazon Machine Learning Developer Guide.

lastUpdatedAt :: Maybe POSIX

The time of the most recent edit to the Evaluation. The time is expressed in epoch time.

createdAt :: Maybe POSIX

The time that the Evaluation was created. The time is expressed in epoch time.

computeTime :: Maybe Integer

The approximate CPU time in milliseconds that Amazon Machine Learning spent processing the Evaluation, normalized and scaled on computation resources. ComputeTime is only available if the Evaluation is in the COMPLETED state.

inputDataLocationS3 :: Maybe Text

The location of the data file or directory in Amazon Simple Storage Service (Amazon S3).

mLModelId :: Maybe Text

The ID of the MLModel that was the focus of the evaluation.

startedAt :: Maybe POSIX

The epoch time when Amazon Machine Learning marked the Evaluation as INPROGRESS. StartedAt isn't available if the Evaluation is in the PENDING state.

finishedAt :: Maybe POSIX

The epoch time when Amazon Machine Learning marked the Evaluation as COMPLETED or FAILED. FinishedAt is only available when the Evaluation is in the COMPLETED or FAILED state.

createdByIamUser :: Maybe Text

The AWS user account that invoked the evaluation. The account type can be either an AWS root account or an AWS Identity and Access Management (IAM) user account.

name :: Maybe Text

A user-supplied name or description of the Evaluation.

logUri :: Maybe Text

A link to the file that contains logs of the CreateEvaluation operation.

evaluationId :: Maybe Text

The evaluation ID which is same as the EvaluationId in the request.

message :: Maybe Text

A description of the most recent details about evaluating the MLModel.

evaluationDataSourceId :: Maybe Text

The DataSource used for this evaluation.

httpStatus :: Int

The response's http status code.

newGetEvaluationResponse

Arguments

:: Int
$sel:httpStatus:GetEvaluationResponse'

-> GetEvaluationResponse

Create a value of GetEvaluationResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:status:GetEvaluationResponse', getEvaluationResponse_status - The status of the evaluation. This element can have one of the following values:

PENDING - Amazon Machine Language (Amazon ML) submitted a request to evaluate an MLModel.

INPROGRESS - The evaluation is underway.

FAILED - The request to evaluate an MLModel did not run to completion. It is not usable.

COMPLETED - The evaluation process completed successfully.

DELETED - The Evaluation is marked as deleted. It is not usable.

$sel:performanceMetrics:GetEvaluationResponse', getEvaluationResponse_performanceMetrics - Measurements of how well the MLModel performed using observations referenced by the DataSource. One of the following metric is returned based on the type of the MLModel:

BinaryAUC: A binary MLModel uses the Area Under the Curve (AUC) technique to measure performance.

RegressionRMSE: A regression MLModel uses the Root Mean Square Error (RMSE) technique to measure performance. RMSE measures the difference between predicted and actual values for a single variable.

MulticlassAvgFScore: A multiclass MLModel uses the F1 score technique to measure performance.

For more information about performance metrics, please see the Amazon Machine Learning Developer Guide.

$sel:lastUpdatedAt:GetEvaluationResponse', getEvaluationResponse_lastUpdatedAt - The time of the most recent edit to the Evaluation. The time is expressed in epoch time.

$sel:createdAt:GetEvaluationResponse', getEvaluationResponse_createdAt - The time that the Evaluation was created. The time is expressed in epoch time.

$sel:computeTime:GetEvaluationResponse', getEvaluationResponse_computeTime - The approximate CPU time in milliseconds that Amazon Machine Learning spent processing the Evaluation, normalized and scaled on computation resources. ComputeTime is only available if the Evaluation is in the COMPLETED state.

$sel:inputDataLocationS3:GetEvaluationResponse', getEvaluationResponse_inputDataLocationS3 - The location of the data file or directory in Amazon Simple Storage Service (Amazon S3).

$sel:mLModelId:GetEvaluationResponse', getEvaluationResponse_mLModelId - The ID of the MLModel that was the focus of the evaluation.

$sel:startedAt:GetEvaluationResponse', getEvaluationResponse_startedAt - The epoch time when Amazon Machine Learning marked the Evaluation as INPROGRESS. StartedAt isn't available if the Evaluation is in the PENDING state.

$sel:finishedAt:GetEvaluationResponse', getEvaluationResponse_finishedAt - The epoch time when Amazon Machine Learning marked the Evaluation as COMPLETED or FAILED. FinishedAt is only available when the Evaluation is in the COMPLETED or FAILED state.

$sel:createdByIamUser:GetEvaluationResponse', getEvaluationResponse_createdByIamUser - The AWS user account that invoked the evaluation. The account type can be either an AWS root account or an AWS Identity and Access Management (IAM) user account.

$sel:name:GetEvaluationResponse', getEvaluationResponse_name - A user-supplied name or description of the Evaluation.

$sel:logUri:GetEvaluationResponse', getEvaluationResponse_logUri - A link to the file that contains logs of the CreateEvaluation operation.

$sel:evaluationId:GetEvaluation', getEvaluationResponse_evaluationId - The evaluation ID which is same as the EvaluationId in the request.

$sel:message:GetEvaluationResponse', getEvaluationResponse_message - A description of the most recent details about evaluating the MLModel.

$sel:evaluationDataSourceId:GetEvaluationResponse', getEvaluationResponse_evaluationDataSourceId - The DataSource used for this evaluation.

$sel:httpStatus:GetEvaluationResponse', getEvaluationResponse_httpStatus - The response's http status code.

Response Lenses

getEvaluationResponse_status :: Lens' GetEvaluationResponse (Maybe EntityStatus)

The status of the evaluation. This element can have one of the following values:

PENDING - Amazon Machine Language (Amazon ML) submitted a request to evaluate an MLModel.

INPROGRESS - The evaluation is underway.

FAILED - The request to evaluate an MLModel did not run to completion. It is not usable.

COMPLETED - The evaluation process completed successfully.

DELETED - The Evaluation is marked as deleted. It is not usable.

getEvaluationResponse_performanceMetrics :: Lens' GetEvaluationResponse (Maybe PerformanceMetrics)

Measurements of how well the MLModel performed using observations referenced by the DataSource. One of the following metric is returned based on the type of the MLModel:

BinaryAUC: A binary MLModel uses the Area Under the Curve (AUC) technique to measure performance.

RegressionRMSE: A regression MLModel uses the Root Mean Square Error (RMSE) technique to measure performance. RMSE measures the difference between predicted and actual values for a single variable.

MulticlassAvgFScore: A multiclass MLModel uses the F1 score technique to measure performance.

For more information about performance metrics, please see the Amazon Machine Learning Developer Guide.

getEvaluationResponse_lastUpdatedAt :: Lens' GetEvaluationResponse (Maybe UTCTime)

The time of the most recent edit to the Evaluation. The time is expressed in epoch time.

getEvaluationResponse_createdAt :: Lens' GetEvaluationResponse (Maybe UTCTime)

The time that the Evaluation was created. The time is expressed in epoch time.

getEvaluationResponse_computeTime :: Lens' GetEvaluationResponse (Maybe Integer)

The approximate CPU time in milliseconds that Amazon Machine Learning spent processing the Evaluation, normalized and scaled on computation resources. ComputeTime is only available if the Evaluation is in the COMPLETED state.

getEvaluationResponse_inputDataLocationS3 :: Lens' GetEvaluationResponse (Maybe Text)

The location of the data file or directory in Amazon Simple Storage Service (Amazon S3).

getEvaluationResponse_mLModelId :: Lens' GetEvaluationResponse (Maybe Text)

The ID of the MLModel that was the focus of the evaluation.

getEvaluationResponse_startedAt :: Lens' GetEvaluationResponse (Maybe UTCTime)

The epoch time when Amazon Machine Learning marked the Evaluation as INPROGRESS. StartedAt isn't available if the Evaluation is in the PENDING state.

getEvaluationResponse_finishedAt :: Lens' GetEvaluationResponse (Maybe UTCTime)

The epoch time when Amazon Machine Learning marked the Evaluation as COMPLETED or FAILED. FinishedAt is only available when the Evaluation is in the COMPLETED or FAILED state.

getEvaluationResponse_createdByIamUser :: Lens' GetEvaluationResponse (Maybe Text)

The AWS user account that invoked the evaluation. The account type can be either an AWS root account or an AWS Identity and Access Management (IAM) user account.

getEvaluationResponse_name :: Lens' GetEvaluationResponse (Maybe Text)

A user-supplied name or description of the Evaluation.

getEvaluationResponse_logUri :: Lens' GetEvaluationResponse (Maybe Text)

A link to the file that contains logs of the CreateEvaluation operation.

getEvaluationResponse_evaluationId :: Lens' GetEvaluationResponse (Maybe Text)

The evaluation ID which is same as the EvaluationId in the request.

getEvaluationResponse_message :: Lens' GetEvaluationResponse (Maybe Text)

A description of the most recent details about evaluating the MLModel.

getEvaluationResponse_evaluationDataSourceId :: Lens' GetEvaluationResponse (Maybe Text)

The DataSource used for this evaluation.

getEvaluationResponse_httpStatus :: Lens' GetEvaluationResponse Int

The response's http status code.