Amazonka.MachineLearning.Types.MLModel

Description

Documentation

data MLModel

Represents the output of a GetMLModel operation.

The content consists of the detailed metadata and the current status of the MLModel.

See: newMLModel smart constructor.

Constructors

MLModel'

Fields

status :: Maybe EntityStatus

The current status of an MLModel. This element can have one of the following values:

PENDING - Amazon Machine Learning (Amazon ML) submitted a request to create an MLModel.

INPROGRESS - The creation process is underway.

FAILED - The request to create an MLModel didn't run to completion. The model isn't usable.

COMPLETED - The creation process completed successfully.

DELETED - The MLModel is marked as deleted. It isn't usable.

lastUpdatedAt :: Maybe POSIX

The time of the most recent edit to the MLModel. The time is expressed in epoch time.

trainingParameters :: Maybe (HashMap Text Text)

A list of the training parameters in the MLModel. The list is implemented as a map of key-value pairs.

The following is the current set of training parameters:

sgd.maxMLModelSizeInBytes - The maximum allowed size of the model. Depending on the input data, the size of the model might affect its performance.

The value is an integer that ranges from 100000 to 2147483648. The default value is 33554432.

sgd.maxPasses - The number of times that the training process traverses the observations to build the MLModel. The value is an integer that ranges from 1 to 10000. The default value is 10.

sgd.shuffleType - Whether Amazon ML shuffles the training data. Shuffling the data improves a model's ability to find the optimal solution for a variety of data types. The valid values are auto and none. The default value is none.

sgd.l1RegularizationAmount - The coefficient regularization L1 norm, which controls overfitting the data by penalizing large coefficients. This parameter tends to drive coefficients to zero, resulting in sparse feature set. If you use this parameter, start by specifying a small value, such as 1.0E-08.

The value is a double that ranges from 0 to MAX_DOUBLE. The default is to not use L1 normalization. This parameter can't be used when L2 is specified. Use this parameter sparingly.

sgd.l2RegularizationAmount - The coefficient regularization L2 norm, which controls overfitting the data by penalizing large coefficients. This tends to drive coefficients to small, nonzero values. If you use this parameter, start by specifying a small value, such as 1.0E-08.

The value is a double that ranges from 0 to MAX_DOUBLE. The default is to not use L2 normalization. This parameter can't be used when L1 is specified. Use this parameter sparingly.

scoreThresholdLastUpdatedAt :: Maybe POSIX

The time of the most recent edit to the ScoreThreshold. The time is expressed in epoch time.

createdAt :: Maybe POSIX

The time that the MLModel was created. The time is expressed in epoch time.

computeTime :: Maybe Integer

inputDataLocationS3 :: Maybe Text

The location of the data file or directory in Amazon Simple Storage Service (Amazon S3).

mLModelId :: Maybe Text

The ID assigned to the MLModel at creation.

sizeInBytes :: Maybe Integer

startedAt :: Maybe POSIX

scoreThreshold :: Maybe Double

finishedAt :: Maybe POSIX

algorithm :: Maybe Algorithm

The algorithm used to train the MLModel. The following algorithm is supported:

SGD -- Stochastic gradient descent. The goal of SGD is to minimize the gradient of the loss function.

createdByIamUser :: Maybe Text

The AWS user account from which the MLModel was created. The account type can be either an AWS root account or an AWS Identity and Access Management (IAM) user account.

name :: Maybe Text

A user-supplied name or description of the MLModel.

endpointInfo :: Maybe RealtimeEndpointInfo

The current endpoint of the MLModel.

trainingDataSourceId :: Maybe Text

The ID of the training DataSource. The CreateMLModel operation uses the TrainingDataSourceId.

message :: Maybe Text

A description of the most recent details about accessing the MLModel.

mLModelType :: Maybe MLModelType

Identifies the MLModel category. The following are the available types:

REGRESSION - Produces a numeric result. For example, "What price should a house be listed at?"

BINARY - Produces one of two possible results. For example, "Is this a child-friendly web site?".

MULTICLASS - Produces one of several possible results. For example, "Is this a HIGH-, LOW-, or MEDIUM-risk trade?".

newMLModel :: MLModel

Create a value of MLModel with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:status:MLModel', mLModel_status - The current status of an MLModel. This element can have one of the following values:

PENDING - Amazon Machine Learning (Amazon ML) submitted a request to create an MLModel.

INPROGRESS - The creation process is underway.

FAILED - The request to create an MLModel didn't run to completion. The model isn't usable.

COMPLETED - The creation process completed successfully.

DELETED - The MLModel is marked as deleted. It isn't usable.

$sel:lastUpdatedAt:MLModel', mLModel_lastUpdatedAt - The time of the most recent edit to the MLModel. The time is expressed in epoch time.

$sel:trainingParameters:MLModel', mLModel_trainingParameters - A list of the training parameters in the MLModel. The list is implemented as a map of key-value pairs.

The following is the current set of training parameters:

sgd.maxMLModelSizeInBytes - The maximum allowed size of the model. Depending on the input data, the size of the model might affect its performance.

The value is an integer that ranges from 100000 to 2147483648. The default value is 33554432.

sgd.maxPasses - The number of times that the training process traverses the observations to build the MLModel. The value is an integer that ranges from 1 to 10000. The default value is 10.

sgd.shuffleType - Whether Amazon ML shuffles the training data. Shuffling the data improves a model's ability to find the optimal solution for a variety of data types. The valid values are auto and none. The default value is none.

sgd.l1RegularizationAmount - The coefficient regularization L1 norm, which controls overfitting the data by penalizing large coefficients. This parameter tends to drive coefficients to zero, resulting in sparse feature set. If you use this parameter, start by specifying a small value, such as 1.0E-08.

The value is a double that ranges from 0 to MAX_DOUBLE. The default is to not use L1 normalization. This parameter can't be used when L2 is specified. Use this parameter sparingly.

sgd.l2RegularizationAmount - The coefficient regularization L2 norm, which controls overfitting the data by penalizing large coefficients. This tends to drive coefficients to small, nonzero values. If you use this parameter, start by specifying a small value, such as 1.0E-08.

The value is a double that ranges from 0 to MAX_DOUBLE. The default is to not use L2 normalization. This parameter can't be used when L1 is specified. Use this parameter sparingly.

$sel:scoreThresholdLastUpdatedAt:MLModel', mLModel_scoreThresholdLastUpdatedAt - The time of the most recent edit to the ScoreThreshold. The time is expressed in epoch time.

$sel:createdAt:MLModel', mLModel_createdAt - The time that the MLModel was created. The time is expressed in epoch time.

$sel:computeTime:MLModel', mLModel_computeTime - Undocumented member.

$sel:inputDataLocationS3:MLModel', mLModel_inputDataLocationS3 - The location of the data file or directory in Amazon Simple Storage Service (Amazon S3).

$sel:mLModelId:MLModel', mLModel_mLModelId - The ID assigned to the MLModel at creation.

$sel:sizeInBytes:MLModel', mLModel_sizeInBytes - Undocumented member.

$sel:startedAt:MLModel', mLModel_startedAt - Undocumented member.

$sel:scoreThreshold:MLModel', mLModel_scoreThreshold - Undocumented member.

$sel:finishedAt:MLModel', mLModel_finishedAt - Undocumented member.

$sel:algorithm:MLModel', mLModel_algorithm - The algorithm used to train the MLModel. The following algorithm is supported:

SGD -- Stochastic gradient descent. The goal of SGD is to minimize the gradient of the loss function.

$sel:createdByIamUser:MLModel', mLModel_createdByIamUser - The AWS user account from which the MLModel was created. The account type can be either an AWS root account or an AWS Identity and Access Management (IAM) user account.

$sel:name:MLModel', mLModel_name - A user-supplied name or description of the MLModel.

$sel:endpointInfo:MLModel', mLModel_endpointInfo - The current endpoint of the MLModel.

$sel:trainingDataSourceId:MLModel', mLModel_trainingDataSourceId - The ID of the training DataSource. The CreateMLModel operation uses the TrainingDataSourceId.

$sel:message:MLModel', mLModel_message - A description of the most recent details about accessing the MLModel.

$sel:mLModelType:MLModel', mLModel_mLModelType - Identifies the MLModel category. The following are the available types:

REGRESSION - Produces a numeric result. For example, "What price should a house be listed at?"

BINARY - Produces one of two possible results. For example, "Is this a child-friendly web site?".

MULTICLASS - Produces one of several possible results. For example, "Is this a HIGH-, LOW-, or MEDIUM-risk trade?".

mLModel_status :: Lens' MLModel (Maybe EntityStatus)

The current status of an MLModel. This element can have one of the following values:

PENDING - Amazon Machine Learning (Amazon ML) submitted a request to create an MLModel.

INPROGRESS - The creation process is underway.

FAILED - The request to create an MLModel didn't run to completion. The model isn't usable.

COMPLETED - The creation process completed successfully.

DELETED - The MLModel is marked as deleted. It isn't usable.

mLModel_lastUpdatedAt :: Lens' MLModel (Maybe UTCTime)

The time of the most recent edit to the MLModel. The time is expressed in epoch time.

mLModel_trainingParameters :: Lens' MLModel (Maybe (HashMap Text Text))

A list of the training parameters in the MLModel. The list is implemented as a map of key-value pairs.

The following is the current set of training parameters:

sgd.maxMLModelSizeInBytes - The maximum allowed size of the model. Depending on the input data, the size of the model might affect its performance.

The value is an integer that ranges from 100000 to 2147483648. The default value is 33554432.

sgd.maxPasses - The number of times that the training process traverses the observations to build the MLModel. The value is an integer that ranges from 1 to 10000. The default value is 10.

sgd.shuffleType - Whether Amazon ML shuffles the training data. Shuffling the data improves a model's ability to find the optimal solution for a variety of data types. The valid values are auto and none. The default value is none.

sgd.l1RegularizationAmount - The coefficient regularization L1 norm, which controls overfitting the data by penalizing large coefficients. This parameter tends to drive coefficients to zero, resulting in sparse feature set. If you use this parameter, start by specifying a small value, such as 1.0E-08.

The value is a double that ranges from 0 to MAX_DOUBLE. The default is to not use L1 normalization. This parameter can't be used when L2 is specified. Use this parameter sparingly.

sgd.l2RegularizationAmount - The coefficient regularization L2 norm, which controls overfitting the data by penalizing large coefficients. This tends to drive coefficients to small, nonzero values. If you use this parameter, start by specifying a small value, such as 1.0E-08.

The value is a double that ranges from 0 to MAX_DOUBLE. The default is to not use L2 normalization. This parameter can't be used when L1 is specified. Use this parameter sparingly.

mLModel_scoreThresholdLastUpdatedAt :: Lens' MLModel (Maybe UTCTime)

The time of the most recent edit to the ScoreThreshold. The time is expressed in epoch time.

mLModel_createdAt :: Lens' MLModel (Maybe UTCTime)

The time that the MLModel was created. The time is expressed in epoch time.

mLModel_computeTime :: Lens' MLModel (Maybe Integer)

Undocumented member.

mLModel_inputDataLocationS3 :: Lens' MLModel (Maybe Text)

The location of the data file or directory in Amazon Simple Storage Service (Amazon S3).

mLModel_mLModelId :: Lens' MLModel (Maybe Text)

The ID assigned to the MLModel at creation.

mLModel_sizeInBytes :: Lens' MLModel (Maybe Integer)

Undocumented member.

mLModel_startedAt :: Lens' MLModel (Maybe UTCTime)

Undocumented member.

mLModel_scoreThreshold :: Lens' MLModel (Maybe Double)

Undocumented member.

mLModel_finishedAt :: Lens' MLModel (Maybe UTCTime)

Undocumented member.

mLModel_algorithm :: Lens' MLModel (Maybe Algorithm)

The algorithm used to train the MLModel. The following algorithm is supported:

SGD -- Stochastic gradient descent. The goal of SGD is to minimize the gradient of the loss function.

mLModel_createdByIamUser :: Lens' MLModel (Maybe Text)

The AWS user account from which the MLModel was created. The account type can be either an AWS root account or an AWS Identity and Access Management (IAM) user account.

mLModel_name :: Lens' MLModel (Maybe Text)

A user-supplied name or description of the MLModel.

mLModel_endpointInfo :: Lens' MLModel (Maybe RealtimeEndpointInfo)

The current endpoint of the MLModel.

mLModel_trainingDataSourceId :: Lens' MLModel (Maybe Text)

The ID of the training DataSource. The CreateMLModel operation uses the TrainingDataSourceId.

mLModel_message :: Lens' MLModel (Maybe Text)

A description of the most recent details about accessing the MLModel.

mLModel_mLModelType :: Lens' MLModel (Maybe MLModelType)

Identifies the MLModel category. The following are the available types:

REGRESSION - Produces a numeric result. For example, "What price should a house be listed at?"

BINARY - Produces one of two possible results. For example, "Is this a child-friendly web site?".

MULTICLASS - Produces one of several possible results. For example, "Is this a HIGH-, LOW-, or MEDIUM-risk trade?".