Amazonka.MachineLearning.UpdateMLModel

Description

Updates the MLModelName and the ScoreThreshold of an MLModel.

You can use the GetMLModel operation to view the contents of the updated data element.

Creating a Request

data UpdateMLModel

See: newUpdateMLModel smart constructor.

Constructors

UpdateMLModel'

Fields

mLModelName :: Maybe Text

A user-supplied name or description of the MLModel.

scoreThreshold :: Maybe Double

The ScoreThreshold used in binary classification MLModel that marks the boundary between a positive prediction and a negative prediction.

Output values greater than or equal to the ScoreThreshold receive a positive result from the MLModel, such as true. Output values less than the ScoreThreshold receive a negative response from the MLModel, such as false.

mLModelId :: Text

The ID assigned to the MLModel during creation.

newUpdateMLModel

Arguments

:: Text
$sel:mLModelId:UpdateMLModel'

-> UpdateMLModel

Create a value of UpdateMLModel with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:mLModelName:UpdateMLModel', updateMLModel_mLModelName - A user-supplied name or description of the MLModel.

$sel:scoreThreshold:UpdateMLModel', updateMLModel_scoreThreshold - The ScoreThreshold used in binary classification MLModel that marks the boundary between a positive prediction and a negative prediction.

Output values greater than or equal to the ScoreThreshold receive a positive result from the MLModel, such as true. Output values less than the ScoreThreshold receive a negative response from the MLModel, such as false.

$sel:mLModelId:UpdateMLModel', updateMLModel_mLModelId - The ID assigned to the MLModel during creation.

Request Lenses

updateMLModel_mLModelName :: Lens' UpdateMLModel (Maybe Text)

A user-supplied name or description of the MLModel.

updateMLModel_scoreThreshold :: Lens' UpdateMLModel (Maybe Double)

The ScoreThreshold used in binary classification MLModel that marks the boundary between a positive prediction and a negative prediction.

Output values greater than or equal to the ScoreThreshold receive a positive result from the MLModel, such as true. Output values less than the ScoreThreshold receive a negative response from the MLModel, such as false.

updateMLModel_mLModelId :: Lens' UpdateMLModel Text

The ID assigned to the MLModel during creation.

Destructuring the Response

data UpdateMLModelResponse

Represents the output of an UpdateMLModel operation.

You can see the updated content by using the GetMLModel operation.

See: newUpdateMLModelResponse smart constructor.

Constructors

UpdateMLModelResponse'

Fields

mLModelId :: Maybe Text

The ID assigned to the MLModel during creation. This value should be identical to the value of the MLModelID in the request.

httpStatus :: Int

The response's http status code.

newUpdateMLModelResponse

Arguments

:: Int
$sel:httpStatus:UpdateMLModelResponse'

-> UpdateMLModelResponse

Create a value of UpdateMLModelResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:mLModelId:UpdateMLModel', updateMLModelResponse_mLModelId - The ID assigned to the MLModel during creation. This value should be identical to the value of the MLModelID in the request.

$sel:httpStatus:UpdateMLModelResponse', updateMLModelResponse_httpStatus - The response's http status code.

Response Lenses

updateMLModelResponse_mLModelId :: Lens' UpdateMLModelResponse (Maybe Text)

The ID assigned to the MLModel during creation. This value should be identical to the value of the MLModelID in the request.

updateMLModelResponse_httpStatus :: Lens' UpdateMLModelResponse Int

The response's http status code.