Amazonka.Macie.Lens

Description

Operations

AssociateS3Resources

associateS3Resources_memberAccountId :: Lens' AssociateS3Resources (Maybe Text)

The ID of the Amazon Macie Classic member account whose resources you want to associate with Macie Classic.

associateS3Resources_s3Resources :: Lens' AssociateS3Resources [S3ResourceClassification]

The S3 resources that you want to associate with Amazon Macie Classic for monitoring and data classification.

associateS3ResourcesResponse_failedS3Resources :: Lens' AssociateS3ResourcesResponse (Maybe [FailedS3Resource])

S3 resources that couldn't be associated with Amazon Macie Classic. An error code and an error message are provided for each failed item.

associateS3ResourcesResponse_httpStatus :: Lens' AssociateS3ResourcesResponse Int

The response's http status code.

AssociateMemberAccount

associateMemberAccount_memberAccountId :: Lens' AssociateMemberAccount Text

The ID of the AWS account that you want to associate with Amazon Macie Classic as a member account.

UpdateS3Resources

updateS3Resources_memberAccountId :: Lens' UpdateS3Resources (Maybe Text)

The AWS ID of the Amazon Macie Classic member account whose S3 resources' classification types you want to update.

updateS3Resources_s3ResourcesUpdate :: Lens' UpdateS3Resources [S3ResourceClassificationUpdate]

The S3 resources whose classification types you want to update.

updateS3ResourcesResponse_failedS3Resources :: Lens' UpdateS3ResourcesResponse (Maybe [FailedS3Resource])

The S3 resources whose classification types can't be updated. An error code and an error message are provided for each failed item.

updateS3ResourcesResponse_httpStatus :: Lens' UpdateS3ResourcesResponse Int

The response's http status code.

ListMemberAccounts

listMemberAccounts_nextToken :: Lens' ListMemberAccounts (Maybe Text)

Use this parameter when paginating results. Set the value of this parameter to null on your first call to the ListMemberAccounts action. Subsequent calls to the action fill nextToken in the request with the value of nextToken from the previous response to continue listing data.

listMemberAccounts_maxResults :: Lens' ListMemberAccounts (Maybe Int)

Use this parameter to indicate the maximum number of items that you want in the response. The default value is 250.

listMemberAccountsResponse_nextToken :: Lens' ListMemberAccountsResponse (Maybe Text)

When a response is generated, if there is more data to be listed, this parameter is present in the response and contains the value to use for the nextToken parameter in a subsequent pagination request. If there is no more data to be listed, this parameter is set to null.

listMemberAccountsResponse_memberAccounts :: Lens' ListMemberAccountsResponse (Maybe [MemberAccount])

A list of the Amazon Macie Classic member accounts returned by the action. The current Macie Classic administrator account is also included in this list.

listMemberAccountsResponse_httpStatus :: Lens' ListMemberAccountsResponse Int

The response's http status code.

DisassociateMemberAccount

disassociateMemberAccount_memberAccountId :: Lens' DisassociateMemberAccount Text

The ID of the member account that you want to remove from Amazon Macie Classic.

ListS3Resources

listS3Resources_memberAccountId :: Lens' ListS3Resources (Maybe Text)

The Amazon Macie Classic member account ID whose associated S3 resources you want to list.

listS3Resources_nextToken :: Lens' ListS3Resources (Maybe Text)

Use this parameter when paginating results. Set its value to null on your first call to the ListS3Resources action. Subsequent calls to the action fill nextToken in the request with the value of nextToken from the previous response to continue listing data.

listS3Resources_maxResults :: Lens' ListS3Resources (Maybe Int)

Use this parameter to indicate the maximum number of items that you want in the response. The default value is 250.

listS3ResourcesResponse_nextToken :: Lens' ListS3ResourcesResponse (Maybe Text)

When a response is generated, if there is more data to be listed, this parameter is present in the response and contains the value to use for the nextToken parameter in a subsequent pagination request. If there is no more data to be listed, this parameter is set to null.

listS3ResourcesResponse_s3Resources :: Lens' ListS3ResourcesResponse (Maybe [S3ResourceClassification])

A list of the associated S3 resources returned by the action.

listS3ResourcesResponse_httpStatus :: Lens' ListS3ResourcesResponse Int

The response's http status code.

DisassociateS3Resources

disassociateS3Resources_memberAccountId :: Lens' DisassociateS3Resources (Maybe Text)

The ID of the Amazon Macie Classic member account whose resources you want to remove from being monitored by Macie Classic.

disassociateS3Resources_associatedS3Resources :: Lens' DisassociateS3Resources [S3Resource]

The S3 resources (buckets or prefixes) that you want to remove from being monitored and classified by Amazon Macie Classic.

disassociateS3ResourcesResponse_failedS3Resources :: Lens' DisassociateS3ResourcesResponse (Maybe [FailedS3Resource])

S3 resources that couldn't be removed from being monitored and classified by Amazon Macie Classic. An error code and an error message are provided for each failed item.

disassociateS3ResourcesResponse_httpStatus :: Lens' DisassociateS3ResourcesResponse Int

The response's http status code.

Types

ClassificationType

classificationType_oneTime :: Lens' ClassificationType S3OneTimeClassificationType

A one-time classification of all of the existing objects in a specified S3 bucket.

classificationType_continuous :: Lens' ClassificationType S3ContinuousClassificationType

A continuous classification of the objects that are added to a specified S3 bucket. Amazon Macie Classic begins performing continuous classification after a bucket is successfully associated with Macie Classic.

ClassificationTypeUpdate

classificationTypeUpdate_continuous :: Lens' ClassificationTypeUpdate (Maybe S3ContinuousClassificationType)

A continuous classification of the objects that are added to a specified S3 bucket. Amazon Macie Classic begins performing continuous classification after a bucket is successfully associated with Macie Classic.

classificationTypeUpdate_oneTime :: Lens' ClassificationTypeUpdate (Maybe S3OneTimeClassificationType)

A one-time classification of all of the existing objects in a specified S3 bucket.

FailedS3Resource

failedS3Resource_errorCode :: Lens' FailedS3Resource (Maybe Text)

The status code of a failed item.

failedS3Resource_errorMessage :: Lens' FailedS3Resource (Maybe Text)

The error message of a failed item.

failedS3Resource_failedItem :: Lens' FailedS3Resource (Maybe S3Resource)

The failed S3 resources.

MemberAccount

memberAccount_accountId :: Lens' MemberAccount (Maybe Text)

The AWS account ID of the Amazon Macie Classic member account.

S3Resource

s3Resource_prefix :: Lens' S3Resource (Maybe Text)

The prefix of the S3 bucket.

s3Resource_bucketName :: Lens' S3Resource Text

The name of the S3 bucket.

S3ResourceClassification

s3ResourceClassification_prefix :: Lens' S3ResourceClassification (Maybe Text)

The prefix of the S3 bucket that you want to associate with Amazon Macie Classic.

s3ResourceClassification_bucketName :: Lens' S3ResourceClassification Text

The name of the S3 bucket that you want to associate with Amazon Macie Classic.

s3ResourceClassification_classificationType :: Lens' S3ResourceClassification ClassificationType

The classification type that you want to specify for the resource associated with Amazon Macie Classic.

S3ResourceClassificationUpdate

s3ResourceClassificationUpdate_prefix :: Lens' S3ResourceClassificationUpdate (Maybe Text)

The prefix of the S3 bucket whose classification types you want to update.

s3ResourceClassificationUpdate_bucketName :: Lens' S3ResourceClassificationUpdate Text

The name of the S3 bucket whose classification types you want to update.

s3ResourceClassificationUpdate_classificationTypeUpdate :: Lens' S3ResourceClassificationUpdate ClassificationTypeUpdate

The classification type that you want to update for the resource associated with Amazon Macie Classic.