Amazonka.Macie.Types

Description

Service Configuration

defaultService :: Service

API version 2017-12-19 of the Amazon Macie SDK configuration.

Errors

_AccessDeniedException :: AsError a => Getting (First ServiceError) a ServiceError

You do not have required permissions to access the requested resource.

_InvalidInputException :: AsError a => Getting (First ServiceError) a ServiceError

The request was rejected because an invalid or out-of-range value was supplied for an input parameter.

_InternalException :: AsError a => Getting (First ServiceError) a ServiceError

Internal server error.

_LimitExceededException :: AsError a => Getting (First ServiceError) a ServiceError

The request was rejected because it attempted to create resources beyond the current AWS account limits. The error code describes the limit exceeded.

S3ContinuousClassificationType

newtype S3ContinuousClassificationType

Constructors

S3ContinuousClassificationType'

Fields

fromS3ContinuousClassificationType :: Text

Bundled Patterns

pattern S3ContinuousClassificationType_FULL :: S3ContinuousClassificationType

S3OneTimeClassificationType

newtype S3OneTimeClassificationType

Constructors

S3OneTimeClassificationType'

Fields

fromS3OneTimeClassificationType :: Text

Bundled Patterns

pattern S3OneTimeClassificationType_FULL :: S3OneTimeClassificationType

pattern S3OneTimeClassificationType_NONE :: S3OneTimeClassificationType

ClassificationType

data ClassificationType

The classification type that Amazon Macie Classic applies to the associated S3 resources.

See: newClassificationType smart constructor.

Constructors

ClassificationType'

Fields

oneTime :: S3OneTimeClassificationType

A one-time classification of all of the existing objects in a specified S3 bucket.

continuous :: S3ContinuousClassificationType

A continuous classification of the objects that are added to a specified S3 bucket. Amazon Macie Classic begins performing continuous classification after a bucket is successfully associated with Macie Classic.

newClassificationType

Arguments

:: S3OneTimeClassificationType
$sel:oneTime:ClassificationType'

-> S3ContinuousClassificationType
$sel:continuous:ClassificationType'

-> ClassificationType

Create a value of ClassificationType with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:oneTime:ClassificationType', classificationType_oneTime - A one-time classification of all of the existing objects in a specified S3 bucket.

$sel:continuous:ClassificationType', classificationType_continuous - A continuous classification of the objects that are added to a specified S3 bucket. Amazon Macie Classic begins performing continuous classification after a bucket is successfully associated with Macie Classic.

classificationType_oneTime :: Lens' ClassificationType S3OneTimeClassificationType

A one-time classification of all of the existing objects in a specified S3 bucket.

classificationType_continuous :: Lens' ClassificationType S3ContinuousClassificationType

A continuous classification of the objects that are added to a specified S3 bucket. Amazon Macie Classic begins performing continuous classification after a bucket is successfully associated with Macie Classic.

ClassificationTypeUpdate

data ClassificationTypeUpdate

The classification type that Amazon Macie Classic applies to the associated S3 resources. At least one of the classification types (oneTime or continuous) must be specified.

See: newClassificationTypeUpdate smart constructor.

Constructors

ClassificationTypeUpdate'

Fields

continuous :: Maybe S3ContinuousClassificationType

A continuous classification of the objects that are added to a specified S3 bucket. Amazon Macie Classic begins performing continuous classification after a bucket is successfully associated with Macie Classic.

oneTime :: Maybe S3OneTimeClassificationType

A one-time classification of all of the existing objects in a specified S3 bucket.

newClassificationTypeUpdate :: ClassificationTypeUpdate

Create a value of ClassificationTypeUpdate with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:continuous:ClassificationTypeUpdate', classificationTypeUpdate_continuous - A continuous classification of the objects that are added to a specified S3 bucket. Amazon Macie Classic begins performing continuous classification after a bucket is successfully associated with Macie Classic.

$sel:oneTime:ClassificationTypeUpdate', classificationTypeUpdate_oneTime - A one-time classification of all of the existing objects in a specified S3 bucket.

classificationTypeUpdate_continuous :: Lens' ClassificationTypeUpdate (Maybe S3ContinuousClassificationType)

A continuous classification of the objects that are added to a specified S3 bucket. Amazon Macie Classic begins performing continuous classification after a bucket is successfully associated with Macie Classic.

classificationTypeUpdate_oneTime :: Lens' ClassificationTypeUpdate (Maybe S3OneTimeClassificationType)

A one-time classification of all of the existing objects in a specified S3 bucket.

FailedS3Resource

data FailedS3Resource

Includes details about the failed S3 resources.

See: newFailedS3Resource smart constructor.

Constructors

FailedS3Resource'

Fields

errorCode :: Maybe Text

The status code of a failed item.

errorMessage :: Maybe Text

The error message of a failed item.

failedItem :: Maybe S3Resource

The failed S3 resources.

newFailedS3Resource :: FailedS3Resource

Create a value of FailedS3Resource with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:errorCode:FailedS3Resource', failedS3Resource_errorCode - The status code of a failed item.

$sel:errorMessage:FailedS3Resource', failedS3Resource_errorMessage - The error message of a failed item.

$sel:failedItem:FailedS3Resource', failedS3Resource_failedItem - The failed S3 resources.

failedS3Resource_errorCode :: Lens' FailedS3Resource (Maybe Text)

The status code of a failed item.

failedS3Resource_errorMessage :: Lens' FailedS3Resource (Maybe Text)

The error message of a failed item.

failedS3Resource_failedItem :: Lens' FailedS3Resource (Maybe S3Resource)

The failed S3 resources.

MemberAccount

data MemberAccount

Contains information about the Amazon Macie Classic member account.

See: newMemberAccount smart constructor.

Constructors

MemberAccount'

Fields

accountId :: Maybe Text

The AWS account ID of the Amazon Macie Classic member account.

newMemberAccount :: MemberAccount

Create a value of MemberAccount with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:accountId:MemberAccount', memberAccount_accountId - The AWS account ID of the Amazon Macie Classic member account.

memberAccount_accountId :: Lens' MemberAccount (Maybe Text)

The AWS account ID of the Amazon Macie Classic member account.

S3Resource

data S3Resource

Contains information about the S3 resource. This data type is used as a request parameter in the DisassociateS3Resources action and can be used as a response parameter in the AssociateS3Resources and UpdateS3Resources actions.

See: newS3Resource smart constructor.

Constructors

S3Resource'

Fields

prefix :: Maybe Text

The prefix of the S3 bucket.

bucketName :: Text

The name of the S3 bucket.

newS3Resource

Arguments

:: Text
$sel:bucketName:S3Resource'

-> S3Resource

Create a value of S3Resource with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:prefix:S3Resource', s3Resource_prefix - The prefix of the S3 bucket.

$sel:bucketName:S3Resource', s3Resource_bucketName - The name of the S3 bucket.

s3Resource_prefix :: Lens' S3Resource (Maybe Text)

The prefix of the S3 bucket.

s3Resource_bucketName :: Lens' S3Resource Text

The name of the S3 bucket.

S3ResourceClassification

data S3ResourceClassification

The S3 resources that you want to associate with Amazon Macie Classic for monitoring and data classification. This data type is used as a request parameter in the AssociateS3Resources action and a response parameter in the ListS3Resources action.

See: newS3ResourceClassification smart constructor.

Constructors

S3ResourceClassification'

Fields

prefix :: Maybe Text

The prefix of the S3 bucket that you want to associate with Amazon Macie Classic.

bucketName :: Text

The name of the S3 bucket that you want to associate with Amazon Macie Classic.

classificationType :: ClassificationType

The classification type that you want to specify for the resource associated with Amazon Macie Classic.

newS3ResourceClassification

Arguments

:: Text
$sel:bucketName:S3ResourceClassification'

-> ClassificationType
$sel:classificationType:S3ResourceClassification'

-> S3ResourceClassification

Create a value of S3ResourceClassification with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:prefix:S3ResourceClassification', s3ResourceClassification_prefix - The prefix of the S3 bucket that you want to associate with Amazon Macie Classic.

$sel:bucketName:S3ResourceClassification', s3ResourceClassification_bucketName - The name of the S3 bucket that you want to associate with Amazon Macie Classic.

$sel:classificationType:S3ResourceClassification', s3ResourceClassification_classificationType - The classification type that you want to specify for the resource associated with Amazon Macie Classic.

s3ResourceClassification_prefix :: Lens' S3ResourceClassification (Maybe Text)

The prefix of the S3 bucket that you want to associate with Amazon Macie Classic.

s3ResourceClassification_bucketName :: Lens' S3ResourceClassification Text

The name of the S3 bucket that you want to associate with Amazon Macie Classic.

s3ResourceClassification_classificationType :: Lens' S3ResourceClassification ClassificationType

The classification type that you want to specify for the resource associated with Amazon Macie Classic.

S3ResourceClassificationUpdate

data S3ResourceClassificationUpdate

The S3 resources whose classification types you want to update. This data type is used as a request parameter in the UpdateS3Resources action.

See: newS3ResourceClassificationUpdate smart constructor.

Constructors

S3ResourceClassificationUpdate'

Fields

prefix :: Maybe Text

The prefix of the S3 bucket whose classification types you want to update.

bucketName :: Text

The name of the S3 bucket whose classification types you want to update.

classificationTypeUpdate :: ClassificationTypeUpdate

The classification type that you want to update for the resource associated with Amazon Macie Classic.

newS3ResourceClassificationUpdate

Arguments

:: Text
$sel:bucketName:S3ResourceClassificationUpdate'

-> ClassificationTypeUpdate
$sel:classificationTypeUpdate:S3ResourceClassificationUpdate'

-> S3ResourceClassificationUpdate

Create a value of S3ResourceClassificationUpdate with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:prefix:S3ResourceClassificationUpdate', s3ResourceClassificationUpdate_prefix - The prefix of the S3 bucket whose classification types you want to update.

$sel:bucketName:S3ResourceClassificationUpdate', s3ResourceClassificationUpdate_bucketName - The name of the S3 bucket whose classification types you want to update.

$sel:classificationTypeUpdate:S3ResourceClassificationUpdate', s3ResourceClassificationUpdate_classificationTypeUpdate - The classification type that you want to update for the resource associated with Amazon Macie Classic.

s3ResourceClassificationUpdate_prefix :: Lens' S3ResourceClassificationUpdate (Maybe Text)

The prefix of the S3 bucket whose classification types you want to update.

s3ResourceClassificationUpdate_bucketName :: Lens' S3ResourceClassificationUpdate Text

The name of the S3 bucket whose classification types you want to update.

s3ResourceClassificationUpdate_classificationTypeUpdate :: Lens' S3ResourceClassificationUpdate ClassificationTypeUpdate

The classification type that you want to update for the resource associated with Amazon Macie Classic.