Amazonka.MacieV2.CreateClassificationJob

Description

Creates and defines the settings for a classification job.

Creating a Request

data CreateClassificationJob

See: newCreateClassificationJob smart constructor.

Constructors

CreateClassificationJob'

Fields

initialRun :: Maybe Bool

For a recurring job, specifies whether to analyze all existing, eligible objects immediately after the job is created (true). To analyze only those objects that are created or changed after you create the job and before the job's first scheduled run, set this value to false.

If you configure the job to run only once, don't specify a value for this property.

samplingPercentage :: Maybe Int

The sampling depth, as a percentage, for the job to apply when processing objects. This value determines the percentage of eligible objects that the job analyzes. If this value is less than 100, Amazon Macie selects the objects to analyze at random, up to the specified percentage, and analyzes all the data in those objects.

managedDataIdentifierSelector :: Maybe ManagedDataIdentifierSelector

The selection type to apply when determining which managed data identifiers the job uses to analyze data. Valid values are:

ALL - Use all the managed data identifiers that Amazon Macie provides. If you specify this value, don't specify any values for the managedDataIdentifierIds property.

EXCLUDE - Use all the managed data identifiers that Macie provides except the managed data identifiers specified by the managedDataIdentifierIds property.

INCLUDE - Use only the managed data identifiers specified by the managedDataIdentifierIds property.

NONE - Don't use any managed data identifiers. If you specify this value, specify at least one custom data identifier for the job (customDataIdentifierIds) and don't specify any values for the managedDataIdentifierIds property.

If you don't specify a value for this property, the job uses all managed data identifiers. If you don't specify a value for this property or you specify ALL or EXCLUDE for a recurring job, the job also uses new managed data identifiers as they are released.

customDataIdentifierIds :: Maybe [Text]

An array of unique identifiers, one for each custom data identifier for the job to use when it analyzes data. To use only managed data identifiers, don't specify a value for this property and specify a value other than NONE for the managedDataIdentifierSelector property.

managedDataIdentifierIds :: Maybe [Text]

An array of unique identifiers, one for each managed data identifier for the job to include (use) or exclude (not use) when it analyzes data. Inclusion or exclusion depends on the managed data identifier selection type that you specify for the job (managedDataIdentifierSelector).

To retrieve a list of valid values for this property, use the ListManagedDataIdentifiers operation.

description :: Maybe Text

A custom description of the job. The description can contain as many as 200 characters.

tags :: Maybe (HashMap Text Text)

A map of key-value pairs that specifies the tags to associate with the job.

A job can have a maximum of 50 tags. Each tag consists of a tag key and an associated tag value. The maximum length of a tag key is 128 characters. The maximum length of a tag value is 256 characters.

scheduleFrequency :: Maybe JobScheduleFrequency

The recurrence pattern for running the job. To run the job only once, don't specify a value for this property and set the value for the jobType property to ONE_TIME.

s3JobDefinition :: S3JobDefinition

The S3 buckets that contain the objects to analyze, and the scope of that analysis.

jobType :: JobType

The schedule for running the job. Valid values are:

ONE_TIME - Run the job only once. If you specify this value, don't specify a value for the scheduleFrequency property.

SCHEDULED - Run the job on a daily, weekly, or monthly basis. If you specify this value, use the scheduleFrequency property to define the recurrence pattern for the job.

clientToken :: Text

A unique, case-sensitive token that you provide to ensure the idempotency of the request.

name :: Text

A custom name for the job. The name can contain as many as 500 characters.

newCreateClassificationJob

Arguments

:: S3JobDefinition
$sel:s3JobDefinition:CreateClassificationJob'

-> JobType
$sel:jobType:CreateClassificationJob'

-> Text
$sel:clientToken:CreateClassificationJob'

-> Text
$sel:name:CreateClassificationJob'

-> CreateClassificationJob

Create a value of CreateClassificationJob with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:initialRun:CreateClassificationJob', createClassificationJob_initialRun - For a recurring job, specifies whether to analyze all existing, eligible objects immediately after the job is created (true). To analyze only those objects that are created or changed after you create the job and before the job's first scheduled run, set this value to false.

If you configure the job to run only once, don't specify a value for this property.

$sel:samplingPercentage:CreateClassificationJob', createClassificationJob_samplingPercentage - The sampling depth, as a percentage, for the job to apply when processing objects. This value determines the percentage of eligible objects that the job analyzes. If this value is less than 100, Amazon Macie selects the objects to analyze at random, up to the specified percentage, and analyzes all the data in those objects.

$sel:managedDataIdentifierSelector:CreateClassificationJob', createClassificationJob_managedDataIdentifierSelector - The selection type to apply when determining which managed data identifiers the job uses to analyze data. Valid values are:

ALL - Use all the managed data identifiers that Amazon Macie provides. If you specify this value, don't specify any values for the managedDataIdentifierIds property.

EXCLUDE - Use all the managed data identifiers that Macie provides except the managed data identifiers specified by the managedDataIdentifierIds property.

INCLUDE - Use only the managed data identifiers specified by the managedDataIdentifierIds property.

NONE - Don't use any managed data identifiers. If you specify this value, specify at least one custom data identifier for the job (customDataIdentifierIds) and don't specify any values for the managedDataIdentifierIds property.

If you don't specify a value for this property, the job uses all managed data identifiers. If you don't specify a value for this property or you specify ALL or EXCLUDE for a recurring job, the job also uses new managed data identifiers as they are released.

$sel:customDataIdentifierIds:CreateClassificationJob', createClassificationJob_customDataIdentifierIds - An array of unique identifiers, one for each custom data identifier for the job to use when it analyzes data. To use only managed data identifiers, don't specify a value for this property and specify a value other than NONE for the managedDataIdentifierSelector property.

$sel:managedDataIdentifierIds:CreateClassificationJob', createClassificationJob_managedDataIdentifierIds - An array of unique identifiers, one for each managed data identifier for the job to include (use) or exclude (not use) when it analyzes data. Inclusion or exclusion depends on the managed data identifier selection type that you specify for the job (managedDataIdentifierSelector).

To retrieve a list of valid values for this property, use the ListManagedDataIdentifiers operation.

$sel:description:CreateClassificationJob', createClassificationJob_description - A custom description of the job. The description can contain as many as 200 characters.

$sel:tags:CreateClassificationJob', createClassificationJob_tags - A map of key-value pairs that specifies the tags to associate with the job.

A job can have a maximum of 50 tags. Each tag consists of a tag key and an associated tag value. The maximum length of a tag key is 128 characters. The maximum length of a tag value is 256 characters.

$sel:scheduleFrequency:CreateClassificationJob', createClassificationJob_scheduleFrequency - The recurrence pattern for running the job. To run the job only once, don't specify a value for this property and set the value for the jobType property to ONE_TIME.

$sel:s3JobDefinition:CreateClassificationJob', createClassificationJob_s3JobDefinition - The S3 buckets that contain the objects to analyze, and the scope of that analysis.

$sel:jobType:CreateClassificationJob', createClassificationJob_jobType - The schedule for running the job. Valid values are:

ONE_TIME - Run the job only once. If you specify this value, don't specify a value for the scheduleFrequency property.

SCHEDULED - Run the job on a daily, weekly, or monthly basis. If you specify this value, use the scheduleFrequency property to define the recurrence pattern for the job.

$sel:clientToken:CreateClassificationJob', createClassificationJob_clientToken - A unique, case-sensitive token that you provide to ensure the idempotency of the request.

$sel:name:CreateClassificationJob', createClassificationJob_name - A custom name for the job. The name can contain as many as 500 characters.

Request Lenses

createClassificationJob_initialRun :: Lens' CreateClassificationJob (Maybe Bool)

For a recurring job, specifies whether to analyze all existing, eligible objects immediately after the job is created (true). To analyze only those objects that are created or changed after you create the job and before the job's first scheduled run, set this value to false.

If you configure the job to run only once, don't specify a value for this property.

createClassificationJob_samplingPercentage :: Lens' CreateClassificationJob (Maybe Int)

The sampling depth, as a percentage, for the job to apply when processing objects. This value determines the percentage of eligible objects that the job analyzes. If this value is less than 100, Amazon Macie selects the objects to analyze at random, up to the specified percentage, and analyzes all the data in those objects.

createClassificationJob_managedDataIdentifierSelector :: Lens' CreateClassificationJob (Maybe ManagedDataIdentifierSelector)

The selection type to apply when determining which managed data identifiers the job uses to analyze data. Valid values are:

ALL - Use all the managed data identifiers that Amazon Macie provides. If you specify this value, don't specify any values for the managedDataIdentifierIds property.

EXCLUDE - Use all the managed data identifiers that Macie provides except the managed data identifiers specified by the managedDataIdentifierIds property.

INCLUDE - Use only the managed data identifiers specified by the managedDataIdentifierIds property.

NONE - Don't use any managed data identifiers. If you specify this value, specify at least one custom data identifier for the job (customDataIdentifierIds) and don't specify any values for the managedDataIdentifierIds property.

If you don't specify a value for this property, the job uses all managed data identifiers. If you don't specify a value for this property or you specify ALL or EXCLUDE for a recurring job, the job also uses new managed data identifiers as they are released.

createClassificationJob_customDataIdentifierIds :: Lens' CreateClassificationJob (Maybe [Text])

An array of unique identifiers, one for each custom data identifier for the job to use when it analyzes data. To use only managed data identifiers, don't specify a value for this property and specify a value other than NONE for the managedDataIdentifierSelector property.

createClassificationJob_managedDataIdentifierIds :: Lens' CreateClassificationJob (Maybe [Text])

An array of unique identifiers, one for each managed data identifier for the job to include (use) or exclude (not use) when it analyzes data. Inclusion or exclusion depends on the managed data identifier selection type that you specify for the job (managedDataIdentifierSelector).

To retrieve a list of valid values for this property, use the ListManagedDataIdentifiers operation.

createClassificationJob_description :: Lens' CreateClassificationJob (Maybe Text)

A custom description of the job. The description can contain as many as 200 characters.

createClassificationJob_tags :: Lens' CreateClassificationJob (Maybe (HashMap Text Text))

A map of key-value pairs that specifies the tags to associate with the job.

A job can have a maximum of 50 tags. Each tag consists of a tag key and an associated tag value. The maximum length of a tag key is 128 characters. The maximum length of a tag value is 256 characters.

createClassificationJob_scheduleFrequency :: Lens' CreateClassificationJob (Maybe JobScheduleFrequency)

The recurrence pattern for running the job. To run the job only once, don't specify a value for this property and set the value for the jobType property to ONE_TIME.

createClassificationJob_s3JobDefinition :: Lens' CreateClassificationJob S3JobDefinition

The S3 buckets that contain the objects to analyze, and the scope of that analysis.

createClassificationJob_jobType :: Lens' CreateClassificationJob JobType

The schedule for running the job. Valid values are:

ONE_TIME - Run the job only once. If you specify this value, don't specify a value for the scheduleFrequency property.

SCHEDULED - Run the job on a daily, weekly, or monthly basis. If you specify this value, use the scheduleFrequency property to define the recurrence pattern for the job.

createClassificationJob_clientToken :: Lens' CreateClassificationJob Text

A unique, case-sensitive token that you provide to ensure the idempotency of the request.

createClassificationJob_name :: Lens' CreateClassificationJob Text

A custom name for the job. The name can contain as many as 500 characters.

Destructuring the Response

data CreateClassificationJobResponse

See: newCreateClassificationJobResponse smart constructor.

Constructors

CreateClassificationJobResponse'

Fields

jobId :: Maybe Text

The unique identifier for the job.

jobArn :: Maybe Text

The Amazon Resource Name (ARN) of the job.

httpStatus :: Int

The response's http status code.

newCreateClassificationJobResponse

Arguments

:: Int
$sel:httpStatus:CreateClassificationJobResponse'

-> CreateClassificationJobResponse

Create a value of CreateClassificationJobResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:jobId:CreateClassificationJobResponse', createClassificationJobResponse_jobId - The unique identifier for the job.

$sel:jobArn:CreateClassificationJobResponse', createClassificationJobResponse_jobArn - The Amazon Resource Name (ARN) of the job.

$sel:httpStatus:CreateClassificationJobResponse', createClassificationJobResponse_httpStatus - The response's http status code.

Response Lenses

createClassificationJobResponse_jobId :: Lens' CreateClassificationJobResponse (Maybe Text)

The unique identifier for the job.

createClassificationJobResponse_jobArn :: Lens' CreateClassificationJobResponse (Maybe Text)

The Amazon Resource Name (ARN) of the job.

createClassificationJobResponse_httpStatus :: Lens' CreateClassificationJobResponse Int

The response's http status code.