Amazonka.MacieV2.CreateCustomDataIdentifier

Description

Creates and defines the criteria and other settings for a custom data identifier.

Creating a Request

data CreateCustomDataIdentifier

See: newCreateCustomDataIdentifier smart constructor.

Constructors

CreateCustomDataIdentifier'

Fields

clientToken :: Maybe Text

A unique, case-sensitive token that you provide to ensure the idempotency of the request.

regex :: Maybe Text

The regular expression (regex) that defines the pattern to match. The expression can contain as many as 512 characters.

name :: Maybe Text

A custom name for the custom data identifier. The name can contain as many as 128 characters.

We strongly recommend that you avoid including any sensitive data in the name of a custom data identifier. Other users of your account might be able to see the identifier's name, depending on the actions that they're allowed to perform in Amazon Macie.

keywords :: Maybe [Text]

An array that lists specific character sequences (keywords), one of which must be within proximity (maximumMatchDistance) of the regular expression to match. The array can contain as many as 50 keywords. Each keyword can contain 3-90 UTF-8 characters. Keywords aren't case sensitive.

ignoreWords :: Maybe [Text]

An array that lists specific character sequences (ignore words) to exclude from the results. If the text matched by the regular expression is the same as any string in this array, Amazon Macie ignores it. The array can contain as many as 10 ignore words. Each ignore word can contain 4-90 UTF-8 characters. Ignore words are case sensitive.

maximumMatchDistance :: Maybe Int

The maximum number of characters that can exist between text that matches the regex pattern and the character sequences specified by the keywords array. Amazon Macie includes or excludes a result based on the proximity of a keyword to text that matches the regex pattern. The distance can be 1-300 characters. The default value is 50.

description :: Maybe Text

A custom description of the custom data identifier. The description can contain as many as 512 characters.

We strongly recommend that you avoid including any sensitive data in the description of a custom data identifier. Other users of your account might be able to see the identifier's description, depending on the actions that they're allowed to perform in Amazon Macie.

tags :: Maybe (HashMap Text Text)

A map of key-value pairs that specifies the tags to associate with the custom data identifier.

A custom data identifier can have a maximum of 50 tags. Each tag consists of a tag key and an associated tag value. The maximum length of a tag key is 128 characters. The maximum length of a tag value is 256 characters.

newCreateCustomDataIdentifier :: CreateCustomDataIdentifier

Create a value of CreateCustomDataIdentifier with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:clientToken:CreateCustomDataIdentifier', createCustomDataIdentifier_clientToken - A unique, case-sensitive token that you provide to ensure the idempotency of the request.

$sel:regex:CreateCustomDataIdentifier', createCustomDataIdentifier_regex - The regular expression (regex) that defines the pattern to match. The expression can contain as many as 512 characters.

$sel:name:CreateCustomDataIdentifier', createCustomDataIdentifier_name - A custom name for the custom data identifier. The name can contain as many as 128 characters.

We strongly recommend that you avoid including any sensitive data in the name of a custom data identifier. Other users of your account might be able to see the identifier's name, depending on the actions that they're allowed to perform in Amazon Macie.

$sel:keywords:CreateCustomDataIdentifier', createCustomDataIdentifier_keywords - An array that lists specific character sequences (keywords), one of which must be within proximity (maximumMatchDistance) of the regular expression to match. The array can contain as many as 50 keywords. Each keyword can contain 3-90 UTF-8 characters. Keywords aren't case sensitive.

$sel:ignoreWords:CreateCustomDataIdentifier', createCustomDataIdentifier_ignoreWords - An array that lists specific character sequences (ignore words) to exclude from the results. If the text matched by the regular expression is the same as any string in this array, Amazon Macie ignores it. The array can contain as many as 10 ignore words. Each ignore word can contain 4-90 UTF-8 characters. Ignore words are case sensitive.

$sel:maximumMatchDistance:CreateCustomDataIdentifier', createCustomDataIdentifier_maximumMatchDistance - The maximum number of characters that can exist between text that matches the regex pattern and the character sequences specified by the keywords array. Amazon Macie includes or excludes a result based on the proximity of a keyword to text that matches the regex pattern. The distance can be 1-300 characters. The default value is 50.

$sel:description:CreateCustomDataIdentifier', createCustomDataIdentifier_description - A custom description of the custom data identifier. The description can contain as many as 512 characters.

We strongly recommend that you avoid including any sensitive data in the description of a custom data identifier. Other users of your account might be able to see the identifier's description, depending on the actions that they're allowed to perform in Amazon Macie.

$sel:tags:CreateCustomDataIdentifier', createCustomDataIdentifier_tags - A map of key-value pairs that specifies the tags to associate with the custom data identifier.

A custom data identifier can have a maximum of 50 tags. Each tag consists of a tag key and an associated tag value. The maximum length of a tag key is 128 characters. The maximum length of a tag value is 256 characters.

Request Lenses

createCustomDataIdentifier_clientToken :: Lens' CreateCustomDataIdentifier (Maybe Text)

A unique, case-sensitive token that you provide to ensure the idempotency of the request.

createCustomDataIdentifier_regex :: Lens' CreateCustomDataIdentifier (Maybe Text)

The regular expression (regex) that defines the pattern to match. The expression can contain as many as 512 characters.

createCustomDataIdentifier_name :: Lens' CreateCustomDataIdentifier (Maybe Text)

A custom name for the custom data identifier. The name can contain as many as 128 characters.

We strongly recommend that you avoid including any sensitive data in the name of a custom data identifier. Other users of your account might be able to see the identifier's name, depending on the actions that they're allowed to perform in Amazon Macie.

createCustomDataIdentifier_keywords :: Lens' CreateCustomDataIdentifier (Maybe [Text])

An array that lists specific character sequences (keywords), one of which must be within proximity (maximumMatchDistance) of the regular expression to match. The array can contain as many as 50 keywords. Each keyword can contain 3-90 UTF-8 characters. Keywords aren't case sensitive.

createCustomDataIdentifier_ignoreWords :: Lens' CreateCustomDataIdentifier (Maybe [Text])

An array that lists specific character sequences (ignore words) to exclude from the results. If the text matched by the regular expression is the same as any string in this array, Amazon Macie ignores it. The array can contain as many as 10 ignore words. Each ignore word can contain 4-90 UTF-8 characters. Ignore words are case sensitive.

createCustomDataIdentifier_maximumMatchDistance :: Lens' CreateCustomDataIdentifier (Maybe Int)

The maximum number of characters that can exist between text that matches the regex pattern and the character sequences specified by the keywords array. Amazon Macie includes or excludes a result based on the proximity of a keyword to text that matches the regex pattern. The distance can be 1-300 characters. The default value is 50.

createCustomDataIdentifier_description :: Lens' CreateCustomDataIdentifier (Maybe Text)

A custom description of the custom data identifier. The description can contain as many as 512 characters.

We strongly recommend that you avoid including any sensitive data in the description of a custom data identifier. Other users of your account might be able to see the identifier's description, depending on the actions that they're allowed to perform in Amazon Macie.

createCustomDataIdentifier_tags :: Lens' CreateCustomDataIdentifier (Maybe (HashMap Text Text))

A map of key-value pairs that specifies the tags to associate with the custom data identifier.

A custom data identifier can have a maximum of 50 tags. Each tag consists of a tag key and an associated tag value. The maximum length of a tag key is 128 characters. The maximum length of a tag value is 256 characters.

Destructuring the Response

data CreateCustomDataIdentifierResponse

See: newCreateCustomDataIdentifierResponse smart constructor.

Constructors

CreateCustomDataIdentifierResponse'

Fields

customDataIdentifierId :: Maybe Text

The unique identifier for the custom data identifier that was created.

httpStatus :: Int

The response's http status code.

newCreateCustomDataIdentifierResponse

Arguments

:: Int
$sel:httpStatus:CreateCustomDataIdentifierResponse'

-> CreateCustomDataIdentifierResponse

Create a value of CreateCustomDataIdentifierResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:customDataIdentifierId:CreateCustomDataIdentifierResponse', createCustomDataIdentifierResponse_customDataIdentifierId - The unique identifier for the custom data identifier that was created.

$sel:httpStatus:CreateCustomDataIdentifierResponse', createCustomDataIdentifierResponse_httpStatus - The response's http status code.

Response Lenses

createCustomDataIdentifierResponse_customDataIdentifierId :: Lens' CreateCustomDataIdentifierResponse (Maybe Text)

The unique identifier for the custom data identifier that was created.

createCustomDataIdentifierResponse_httpStatus :: Lens' CreateCustomDataIdentifierResponse Int

The response's http status code.