Amazonka.MacieV2.CreateFindingsFilter

Description

Creates and defines the criteria and other settings for a findings filter.

Creating a Request

data CreateFindingsFilter

See: newCreateFindingsFilter smart constructor.

Constructors

CreateFindingsFilter'

Fields

clientToken :: Maybe Text

A unique, case-sensitive token that you provide to ensure the idempotency of the request.

description :: Maybe Text

A custom description of the filter. The description can contain as many as 512 characters.

We strongly recommend that you avoid including any sensitive data in the description of a filter. Other users of your account might be able to see the filter's description, depending on the actions that they're allowed to perform in Amazon Macie.

tags :: Maybe (HashMap Text Text)

A map of key-value pairs that specifies the tags to associate with the filter.

A findings filter can have a maximum of 50 tags. Each tag consists of a tag key and an associated tag value. The maximum length of a tag key is 128 characters. The maximum length of a tag value is 256 characters.

position :: Maybe Int

The position of the filter in the list of saved filters on the Amazon Macie console. This value also determines the order in which the filter is applied to findings, relative to other filters that are also applied to the findings.

action :: FindingsFilterAction

The action to perform on findings that meet the filter criteria (findingCriteria). Valid values are: ARCHIVE, suppress (automatically archive) the findings; and, NOOP, don't perform any action on the findings.

findingCriteria :: FindingCriteria

The criteria to use to filter findings.

name :: Text

A custom name for the filter. The name must contain at least 3 characters and can contain as many as 64 characters.

We strongly recommend that you avoid including any sensitive data in the name of a filter. Other users of your account might be able to see the filter's name, depending on the actions that they're allowed to perform in Amazon Macie.

newCreateFindingsFilter

Arguments

:: FindingsFilterAction
$sel:action:CreateFindingsFilter'

-> FindingCriteria
$sel:findingCriteria:CreateFindingsFilter'

-> Text
$sel:name:CreateFindingsFilter'

-> CreateFindingsFilter

Create a value of CreateFindingsFilter with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:clientToken:CreateFindingsFilter', createFindingsFilter_clientToken - A unique, case-sensitive token that you provide to ensure the idempotency of the request.

$sel:description:CreateFindingsFilter', createFindingsFilter_description - A custom description of the filter. The description can contain as many as 512 characters.

We strongly recommend that you avoid including any sensitive data in the description of a filter. Other users of your account might be able to see the filter's description, depending on the actions that they're allowed to perform in Amazon Macie.

$sel:tags:CreateFindingsFilter', createFindingsFilter_tags - A map of key-value pairs that specifies the tags to associate with the filter.

A findings filter can have a maximum of 50 tags. Each tag consists of a tag key and an associated tag value. The maximum length of a tag key is 128 characters. The maximum length of a tag value is 256 characters.

$sel:position:CreateFindingsFilter', createFindingsFilter_position - The position of the filter in the list of saved filters on the Amazon Macie console. This value also determines the order in which the filter is applied to findings, relative to other filters that are also applied to the findings.

$sel:action:CreateFindingsFilter', createFindingsFilter_action - The action to perform on findings that meet the filter criteria (findingCriteria). Valid values are: ARCHIVE, suppress (automatically archive) the findings; and, NOOP, don't perform any action on the findings.

$sel:findingCriteria:CreateFindingsFilter', createFindingsFilter_findingCriteria - The criteria to use to filter findings.

$sel:name:CreateFindingsFilter', createFindingsFilter_name - A custom name for the filter. The name must contain at least 3 characters and can contain as many as 64 characters.

We strongly recommend that you avoid including any sensitive data in the name of a filter. Other users of your account might be able to see the filter's name, depending on the actions that they're allowed to perform in Amazon Macie.

Request Lenses

createFindingsFilter_clientToken :: Lens' CreateFindingsFilter (Maybe Text)

A unique, case-sensitive token that you provide to ensure the idempotency of the request.

createFindingsFilter_description :: Lens' CreateFindingsFilter (Maybe Text)

A custom description of the filter. The description can contain as many as 512 characters.

We strongly recommend that you avoid including any sensitive data in the description of a filter. Other users of your account might be able to see the filter's description, depending on the actions that they're allowed to perform in Amazon Macie.

createFindingsFilter_tags :: Lens' CreateFindingsFilter (Maybe (HashMap Text Text))

A map of key-value pairs that specifies the tags to associate with the filter.

A findings filter can have a maximum of 50 tags. Each tag consists of a tag key and an associated tag value. The maximum length of a tag key is 128 characters. The maximum length of a tag value is 256 characters.

createFindingsFilter_position :: Lens' CreateFindingsFilter (Maybe Int)

The position of the filter in the list of saved filters on the Amazon Macie console. This value also determines the order in which the filter is applied to findings, relative to other filters that are also applied to the findings.

createFindingsFilter_action :: Lens' CreateFindingsFilter FindingsFilterAction

The action to perform on findings that meet the filter criteria (findingCriteria). Valid values are: ARCHIVE, suppress (automatically archive) the findings; and, NOOP, don't perform any action on the findings.

createFindingsFilter_findingCriteria :: Lens' CreateFindingsFilter FindingCriteria

The criteria to use to filter findings.

createFindingsFilter_name :: Lens' CreateFindingsFilter Text

A custom name for the filter. The name must contain at least 3 characters and can contain as many as 64 characters.

We strongly recommend that you avoid including any sensitive data in the name of a filter. Other users of your account might be able to see the filter's name, depending on the actions that they're allowed to perform in Amazon Macie.

Destructuring the Response

data CreateFindingsFilterResponse

See: newCreateFindingsFilterResponse smart constructor.

Constructors

CreateFindingsFilterResponse'

Fields

arn :: Maybe Text

The Amazon Resource Name (ARN) of the filter that was created.

id :: Maybe Text

The unique identifier for the filter that was created.

httpStatus :: Int

The response's http status code.

newCreateFindingsFilterResponse

Arguments

:: Int
$sel:httpStatus:CreateFindingsFilterResponse'

-> CreateFindingsFilterResponse

Create a value of CreateFindingsFilterResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:arn:CreateFindingsFilterResponse', createFindingsFilterResponse_arn - The Amazon Resource Name (ARN) of the filter that was created.

$sel:id:CreateFindingsFilterResponse', createFindingsFilterResponse_id - The unique identifier for the filter that was created.

$sel:httpStatus:CreateFindingsFilterResponse', createFindingsFilterResponse_httpStatus - The response's http status code.

Response Lenses

createFindingsFilterResponse_arn :: Lens' CreateFindingsFilterResponse (Maybe Text)

The Amazon Resource Name (ARN) of the filter that was created.

createFindingsFilterResponse_id :: Lens' CreateFindingsFilterResponse (Maybe Text)

The unique identifier for the filter that was created.

createFindingsFilterResponse_httpStatus :: Lens' CreateFindingsFilterResponse Int

The response's http status code.