Amazonka.MacieV2.DescribeBuckets

Description

Retrieves (queries) statistical data and other information about one or more S3 buckets that Amazon Macie monitors and analyzes.

This operation returns paginated results.

Creating a Request

data DescribeBuckets

See: newDescribeBuckets smart constructor.

Constructors

DescribeBuckets'

Fields

sortCriteria :: Maybe BucketSortCriteria

The criteria to use to sort the query results.

nextToken :: Maybe Text

The nextToken string that specifies which page of results to return in a paginated response.

criteria :: Maybe (HashMap Text BucketCriteriaAdditionalProperties)

The criteria to use to filter the query results.

maxResults :: Maybe Int

The maximum number of items to include in each page of the response. The default value is 50.

newDescribeBuckets :: DescribeBuckets

Create a value of DescribeBuckets with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:sortCriteria:DescribeBuckets', describeBuckets_sortCriteria - The criteria to use to sort the query results.

$sel:nextToken:DescribeBuckets', describeBuckets_nextToken - The nextToken string that specifies which page of results to return in a paginated response.

$sel:criteria:DescribeBuckets', describeBuckets_criteria - The criteria to use to filter the query results.

$sel:maxResults:DescribeBuckets', describeBuckets_maxResults - The maximum number of items to include in each page of the response. The default value is 50.

Request Lenses

describeBuckets_sortCriteria :: Lens' DescribeBuckets (Maybe BucketSortCriteria)

The criteria to use to sort the query results.

describeBuckets_nextToken :: Lens' DescribeBuckets (Maybe Text)

The nextToken string that specifies which page of results to return in a paginated response.

describeBuckets_criteria :: Lens' DescribeBuckets (Maybe (HashMap Text BucketCriteriaAdditionalProperties))

The criteria to use to filter the query results.

describeBuckets_maxResults :: Lens' DescribeBuckets (Maybe Int)

The maximum number of items to include in each page of the response. The default value is 50.

Destructuring the Response

data DescribeBucketsResponse

See: newDescribeBucketsResponse smart constructor.

Constructors

DescribeBucketsResponse'

Fields

buckets :: Maybe [BucketMetadata]

An array of objects, one for each bucket that meets the filter criteria specified in the request.

nextToken :: Maybe Text

The string to use in a subsequent request to get the next page of results in a paginated response. This value is null if there are no additional pages.

httpStatus :: Int

The response's http status code.

newDescribeBucketsResponse

Arguments

:: Int
$sel:httpStatus:DescribeBucketsResponse'

-> DescribeBucketsResponse

Create a value of DescribeBucketsResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:buckets:DescribeBucketsResponse', describeBucketsResponse_buckets - An array of objects, one for each bucket that meets the filter criteria specified in the request.

$sel:nextToken:DescribeBuckets', describeBucketsResponse_nextToken - The string to use in a subsequent request to get the next page of results in a paginated response. This value is null if there are no additional pages.

$sel:httpStatus:DescribeBucketsResponse', describeBucketsResponse_httpStatus - The response's http status code.

Response Lenses

describeBucketsResponse_buckets :: Lens' DescribeBucketsResponse (Maybe [BucketMetadata])

An array of objects, one for each bucket that meets the filter criteria specified in the request.

describeBucketsResponse_nextToken :: Lens' DescribeBucketsResponse (Maybe Text)

The string to use in a subsequent request to get the next page of results in a paginated response. This value is null if there are no additional pages.

describeBucketsResponse_httpStatus :: Lens' DescribeBucketsResponse Int

The response's http status code.