Amazonka.MacieV2.DescribeClassificationJob

Description

Retrieves the status and settings for a classification job.

Creating a Request

data DescribeClassificationJob

See: newDescribeClassificationJob smart constructor.

Constructors

DescribeClassificationJob'

Fields

jobId :: Text

The unique identifier for the classification job.

newDescribeClassificationJob

Arguments

:: Text
$sel:jobId:DescribeClassificationJob'

-> DescribeClassificationJob

Create a value of DescribeClassificationJob with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:jobId:DescribeClassificationJob', describeClassificationJob_jobId - The unique identifier for the classification job.

Request Lenses

describeClassificationJob_jobId :: Lens' DescribeClassificationJob Text

The unique identifier for the classification job.

Destructuring the Response

data DescribeClassificationJobResponse

See: newDescribeClassificationJobResponse smart constructor.

Constructors

DescribeClassificationJobResponse'

Fields

lastRunErrorStatus :: Maybe LastRunErrorStatus

Specifies whether any account- or bucket-level access errors occurred when the job ran. For a recurring job, this value indicates the error status of the job's most recent run.

jobType :: Maybe JobType

The schedule for running the job. Possible values are:

ONE_TIME - The job runs only once.

SCHEDULED - The job runs on a daily, weekly, or monthly basis. The scheduleFrequency property indicates the recurrence pattern for the job.

initialRun :: Maybe Bool

For a recurring job, specifies whether you configured the job to analyze all existing, eligible objects immediately after the job was created (true). If you configured the job to analyze only those objects that were created or changed after the job was created and before the job's first scheduled run, this value is false. This value is also false for a one-time job.

jobId :: Maybe Text

The unique identifier for the job.

clientToken :: Maybe Text

The token that was provided to ensure the idempotency of the request to create the job.

jobArn :: Maybe Text

The Amazon Resource Name (ARN) of the job.

s3JobDefinition :: Maybe S3JobDefinition

The S3 buckets that contain the objects to analyze, and the scope of that analysis.

createdAt :: Maybe POSIX

The date and time, in UTC and extended ISO 8601 format, when the job was created.

userPausedDetails :: Maybe UserPausedDetails

If the current status of the job is USER_PAUSED, specifies when the job was paused and when the job or job run will expire and be cancelled if it isn't resumed. This value is present only if the value for jobStatus is USER_PAUSED.

samplingPercentage :: Maybe Int

The sampling depth, as a percentage, that determines the percentage of eligible objects that the job analyzes.

managedDataIdentifierSelector :: Maybe ManagedDataIdentifierSelector

The selection type that determines which managed data identifiers the job uses to analyze data. Possible values are:

ALL - Use all the managed data identifiers that Amazon Macie provides.

EXCLUDE - Use all the managed data identifiers that Macie provides except the managed data identifiers specified by the managedDataIdentifierIds property.

INCLUDE - Use only the managed data identifiers specified by the managedDataIdentifierIds property.

NONE - Don't use any managed data identifiers.

If this value is null, the job uses all managed data identifiers. If this value is null, ALL, or EXCLUDE for a recurring job, the job also uses new managed data identifiers as they are released.

lastRunTime :: Maybe POSIX

The date and time, in UTC and extended ISO 8601 format, when the job started. If the job is a recurring job, this value indicates when the most recent run started.

customDataIdentifierIds :: Maybe [Text]

An array of unique identifiers, one for each custom data identifier that the job uses to analyze data. This value is null if the job uses only managed data identifiers to analyze data.

name :: Maybe Text

The custom name of the job.

statistics :: Maybe Statistics

The number of times that the job has run and processing statistics for the job's current run.

managedDataIdentifierIds :: Maybe [Text]

An array of unique identifiers, one for each managed data identifier that the job is explicitly configured to include (use) or exclude (not use) when it analyzes data. Inclusion or exclusion depends on the managed data identifier selection type specified for the job (managedDataIdentifierSelector). This value is null if the job's managed data identifier selection type is ALL or the job uses only custom data identifiers (customDataIdentifierIds) to analyze data.

jobStatus :: Maybe JobStatus

The current status of the job. Possible values are:

CANCELLED - You cancelled the job or, if it's a one-time job, you paused the job and didn't resume it within 30 days.

COMPLETE - For a one-time job, Amazon Macie finished processing the data specified for the job. This value doesn't apply to recurring jobs.

IDLE - For a recurring job, the previous scheduled run is complete and the next scheduled run is pending. This value doesn't apply to one-time jobs.

PAUSED - Macie started running the job but additional processing would exceed the monthly sensitive data discovery quota for your account or one or more member accounts that the job analyzes data for.

RUNNING - For a one-time job, the job is in progress. For a recurring job, a scheduled run is in progress.

USER_PAUSED - You paused the job. If you paused the job while it had a status of RUNNING and you don't resume it within 30 days of pausing it, the job or job run will expire and be cancelled, depending on the job's type. To check the expiration date, refer to the UserPausedDetails.jobExpiresAt property.

description :: Maybe Text

The custom description of the job.

tags :: Maybe (HashMap Text Text)

A map of key-value pairs that specifies which tags (keys and values) are associated with the classification job.

scheduleFrequency :: Maybe JobScheduleFrequency

The recurrence pattern for running the job. This value is null if the job is configured to run only once.

httpStatus :: Int

The response's http status code.

newDescribeClassificationJobResponse

Arguments

:: Int
$sel:httpStatus:DescribeClassificationJobResponse'

-> DescribeClassificationJobResponse

Create a value of DescribeClassificationJobResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:lastRunErrorStatus:DescribeClassificationJobResponse', describeClassificationJobResponse_lastRunErrorStatus - Specifies whether any account- or bucket-level access errors occurred when the job ran. For a recurring job, this value indicates the error status of the job's most recent run.

$sel:jobType:DescribeClassificationJobResponse', describeClassificationJobResponse_jobType - The schedule for running the job. Possible values are:

ONE_TIME - The job runs only once.

SCHEDULED - The job runs on a daily, weekly, or monthly basis. The scheduleFrequency property indicates the recurrence pattern for the job.

$sel:initialRun:DescribeClassificationJobResponse', describeClassificationJobResponse_initialRun - For a recurring job, specifies whether you configured the job to analyze all existing, eligible objects immediately after the job was created (true). If you configured the job to analyze only those objects that were created or changed after the job was created and before the job's first scheduled run, this value is false. This value is also false for a one-time job.

$sel:jobId:DescribeClassificationJob', describeClassificationJobResponse_jobId - The unique identifier for the job.

$sel:clientToken:DescribeClassificationJobResponse', describeClassificationJobResponse_clientToken - The token that was provided to ensure the idempotency of the request to create the job.

$sel:jobArn:DescribeClassificationJobResponse', describeClassificationJobResponse_jobArn - The Amazon Resource Name (ARN) of the job.

$sel:s3JobDefinition:DescribeClassificationJobResponse', describeClassificationJobResponse_s3JobDefinition - The S3 buckets that contain the objects to analyze, and the scope of that analysis.

$sel:createdAt:DescribeClassificationJobResponse', describeClassificationJobResponse_createdAt - The date and time, in UTC and extended ISO 8601 format, when the job was created.

$sel:userPausedDetails:DescribeClassificationJobResponse', describeClassificationJobResponse_userPausedDetails - If the current status of the job is USER_PAUSED, specifies when the job was paused and when the job or job run will expire and be cancelled if it isn't resumed. This value is present only if the value for jobStatus is USER_PAUSED.

$sel:samplingPercentage:DescribeClassificationJobResponse', describeClassificationJobResponse_samplingPercentage - The sampling depth, as a percentage, that determines the percentage of eligible objects that the job analyzes.

$sel:managedDataIdentifierSelector:DescribeClassificationJobResponse', describeClassificationJobResponse_managedDataIdentifierSelector - The selection type that determines which managed data identifiers the job uses to analyze data. Possible values are:

ALL - Use all the managed data identifiers that Amazon Macie provides.

EXCLUDE - Use all the managed data identifiers that Macie provides except the managed data identifiers specified by the managedDataIdentifierIds property.

INCLUDE - Use only the managed data identifiers specified by the managedDataIdentifierIds property.

NONE - Don't use any managed data identifiers.

If this value is null, the job uses all managed data identifiers. If this value is null, ALL, or EXCLUDE for a recurring job, the job also uses new managed data identifiers as they are released.

$sel:lastRunTime:DescribeClassificationJobResponse', describeClassificationJobResponse_lastRunTime - The date and time, in UTC and extended ISO 8601 format, when the job started. If the job is a recurring job, this value indicates when the most recent run started.

$sel:customDataIdentifierIds:DescribeClassificationJobResponse', describeClassificationJobResponse_customDataIdentifierIds - An array of unique identifiers, one for each custom data identifier that the job uses to analyze data. This value is null if the job uses only managed data identifiers to analyze data.

$sel:name:DescribeClassificationJobResponse', describeClassificationJobResponse_name - The custom name of the job.

$sel:statistics:DescribeClassificationJobResponse', describeClassificationJobResponse_statistics - The number of times that the job has run and processing statistics for the job's current run.

$sel:managedDataIdentifierIds:DescribeClassificationJobResponse', describeClassificationJobResponse_managedDataIdentifierIds - An array of unique identifiers, one for each managed data identifier that the job is explicitly configured to include (use) or exclude (not use) when it analyzes data. Inclusion or exclusion depends on the managed data identifier selection type specified for the job (managedDataIdentifierSelector). This value is null if the job's managed data identifier selection type is ALL or the job uses only custom data identifiers (customDataIdentifierIds) to analyze data.

$sel:jobStatus:DescribeClassificationJobResponse', describeClassificationJobResponse_jobStatus - The current status of the job. Possible values are:

CANCELLED - You cancelled the job or, if it's a one-time job, you paused the job and didn't resume it within 30 days.

COMPLETE - For a one-time job, Amazon Macie finished processing the data specified for the job. This value doesn't apply to recurring jobs.

IDLE - For a recurring job, the previous scheduled run is complete and the next scheduled run is pending. This value doesn't apply to one-time jobs.

PAUSED - Macie started running the job but additional processing would exceed the monthly sensitive data discovery quota for your account or one or more member accounts that the job analyzes data for.

RUNNING - For a one-time job, the job is in progress. For a recurring job, a scheduled run is in progress.

USER_PAUSED - You paused the job. If you paused the job while it had a status of RUNNING and you don't resume it within 30 days of pausing it, the job or job run will expire and be cancelled, depending on the job's type. To check the expiration date, refer to the UserPausedDetails.jobExpiresAt property.

$sel:description:DescribeClassificationJobResponse', describeClassificationJobResponse_description - The custom description of the job.

$sel:tags:DescribeClassificationJobResponse', describeClassificationJobResponse_tags - A map of key-value pairs that specifies which tags (keys and values) are associated with the classification job.

$sel:scheduleFrequency:DescribeClassificationJobResponse', describeClassificationJobResponse_scheduleFrequency - The recurrence pattern for running the job. This value is null if the job is configured to run only once.

$sel:httpStatus:DescribeClassificationJobResponse', describeClassificationJobResponse_httpStatus - The response's http status code.

Response Lenses

describeClassificationJobResponse_lastRunErrorStatus :: Lens' DescribeClassificationJobResponse (Maybe LastRunErrorStatus)

Specifies whether any account- or bucket-level access errors occurred when the job ran. For a recurring job, this value indicates the error status of the job's most recent run.

describeClassificationJobResponse_jobType :: Lens' DescribeClassificationJobResponse (Maybe JobType)

The schedule for running the job. Possible values are:

ONE_TIME - The job runs only once.

SCHEDULED - The job runs on a daily, weekly, or monthly basis. The scheduleFrequency property indicates the recurrence pattern for the job.

describeClassificationJobResponse_initialRun :: Lens' DescribeClassificationJobResponse (Maybe Bool)

For a recurring job, specifies whether you configured the job to analyze all existing, eligible objects immediately after the job was created (true). If you configured the job to analyze only those objects that were created or changed after the job was created and before the job's first scheduled run, this value is false. This value is also false for a one-time job.

describeClassificationJobResponse_jobId :: Lens' DescribeClassificationJobResponse (Maybe Text)

The unique identifier for the job.

describeClassificationJobResponse_clientToken :: Lens' DescribeClassificationJobResponse (Maybe Text)

The token that was provided to ensure the idempotency of the request to create the job.

describeClassificationJobResponse_jobArn :: Lens' DescribeClassificationJobResponse (Maybe Text)

The Amazon Resource Name (ARN) of the job.

describeClassificationJobResponse_s3JobDefinition :: Lens' DescribeClassificationJobResponse (Maybe S3JobDefinition)

The S3 buckets that contain the objects to analyze, and the scope of that analysis.

describeClassificationJobResponse_createdAt :: Lens' DescribeClassificationJobResponse (Maybe UTCTime)

The date and time, in UTC and extended ISO 8601 format, when the job was created.

describeClassificationJobResponse_userPausedDetails :: Lens' DescribeClassificationJobResponse (Maybe UserPausedDetails)

If the current status of the job is USER_PAUSED, specifies when the job was paused and when the job or job run will expire and be cancelled if it isn't resumed. This value is present only if the value for jobStatus is USER_PAUSED.

describeClassificationJobResponse_samplingPercentage :: Lens' DescribeClassificationJobResponse (Maybe Int)

The sampling depth, as a percentage, that determines the percentage of eligible objects that the job analyzes.

describeClassificationJobResponse_managedDataIdentifierSelector :: Lens' DescribeClassificationJobResponse (Maybe ManagedDataIdentifierSelector)

The selection type that determines which managed data identifiers the job uses to analyze data. Possible values are:

ALL - Use all the managed data identifiers that Amazon Macie provides.

EXCLUDE - Use all the managed data identifiers that Macie provides except the managed data identifiers specified by the managedDataIdentifierIds property.

INCLUDE - Use only the managed data identifiers specified by the managedDataIdentifierIds property.

NONE - Don't use any managed data identifiers.

If this value is null, the job uses all managed data identifiers. If this value is null, ALL, or EXCLUDE for a recurring job, the job also uses new managed data identifiers as they are released.

describeClassificationJobResponse_lastRunTime :: Lens' DescribeClassificationJobResponse (Maybe UTCTime)

The date and time, in UTC and extended ISO 8601 format, when the job started. If the job is a recurring job, this value indicates when the most recent run started.

describeClassificationJobResponse_customDataIdentifierIds :: Lens' DescribeClassificationJobResponse (Maybe [Text])

An array of unique identifiers, one for each custom data identifier that the job uses to analyze data. This value is null if the job uses only managed data identifiers to analyze data.

describeClassificationJobResponse_name :: Lens' DescribeClassificationJobResponse (Maybe Text)

The custom name of the job.

describeClassificationJobResponse_statistics :: Lens' DescribeClassificationJobResponse (Maybe Statistics)

The number of times that the job has run and processing statistics for the job's current run.

describeClassificationJobResponse_managedDataIdentifierIds :: Lens' DescribeClassificationJobResponse (Maybe [Text])

An array of unique identifiers, one for each managed data identifier that the job is explicitly configured to include (use) or exclude (not use) when it analyzes data. Inclusion or exclusion depends on the managed data identifier selection type specified for the job (managedDataIdentifierSelector). This value is null if the job's managed data identifier selection type is ALL or the job uses only custom data identifiers (customDataIdentifierIds) to analyze data.

describeClassificationJobResponse_jobStatus :: Lens' DescribeClassificationJobResponse (Maybe JobStatus)

The current status of the job. Possible values are:

CANCELLED - You cancelled the job or, if it's a one-time job, you paused the job and didn't resume it within 30 days.

COMPLETE - For a one-time job, Amazon Macie finished processing the data specified for the job. This value doesn't apply to recurring jobs.

IDLE - For a recurring job, the previous scheduled run is complete and the next scheduled run is pending. This value doesn't apply to one-time jobs.

PAUSED - Macie started running the job but additional processing would exceed the monthly sensitive data discovery quota for your account or one or more member accounts that the job analyzes data for.

RUNNING - For a one-time job, the job is in progress. For a recurring job, a scheduled run is in progress.

USER_PAUSED - You paused the job. If you paused the job while it had a status of RUNNING and you don't resume it within 30 days of pausing it, the job or job run will expire and be cancelled, depending on the job's type. To check the expiration date, refer to the UserPausedDetails.jobExpiresAt property.

describeClassificationJobResponse_description :: Lens' DescribeClassificationJobResponse (Maybe Text)

The custom description of the job.

describeClassificationJobResponse_tags :: Lens' DescribeClassificationJobResponse (Maybe (HashMap Text Text))

A map of key-value pairs that specifies which tags (keys and values) are associated with the classification job.

describeClassificationJobResponse_scheduleFrequency :: Lens' DescribeClassificationJobResponse (Maybe JobScheduleFrequency)

The recurrence pattern for running the job. This value is null if the job is configured to run only once.

describeClassificationJobResponse_httpStatus :: Lens' DescribeClassificationJobResponse Int

The response's http status code.