Amazonka.MacieV2.DisableMacie

Description

Disables an Amazon Macie account and deletes Macie resources for the account.

Creating a Request

data DisableMacie

See: newDisableMacie smart constructor.

Constructors

DisableMacie'

newDisableMacie :: DisableMacie

Create a value of DisableMacie with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

Destructuring the Response

data DisableMacieResponse

See: newDisableMacieResponse smart constructor.

Constructors

DisableMacieResponse'

Fields

httpStatus :: Int

The response's http status code.

newDisableMacieResponse

Arguments

:: Int
$sel:httpStatus:DisableMacieResponse'

-> DisableMacieResponse

Create a value of DisableMacieResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:httpStatus:DisableMacieResponse', disableMacieResponse_httpStatus - The response's http status code.

Response Lenses

disableMacieResponse_httpStatus :: Lens' DisableMacieResponse Int

The response's http status code.