Amazonka.MacieV2.GetBucketStatistics

Description

Retrieves (queries) aggregated statistical data for all the S3 buckets that Amazon Macie monitors and analyzes.

Creating a Request

data GetBucketStatistics

See: newGetBucketStatistics smart constructor.

Constructors

GetBucketStatistics'

Fields

accountId :: Maybe Text

The unique identifier for the Amazon Web Services account.

newGetBucketStatistics :: GetBucketStatistics

Create a value of GetBucketStatistics with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:accountId:GetBucketStatistics', getBucketStatistics_accountId - The unique identifier for the Amazon Web Services account.

Request Lenses

getBucketStatistics_accountId :: Lens' GetBucketStatistics (Maybe Text)

The unique identifier for the Amazon Web Services account.

Destructuring the Response

data GetBucketStatisticsResponse

See: newGetBucketStatisticsResponse smart constructor.

Constructors

GetBucketStatisticsResponse'

Fields

sizeInBytesCompressed :: Maybe Integer

The total storage size, in bytes, of the objects that are compressed (.gz, .gzip, .zip) files in the buckets.

If versioning is enabled for any of the buckets, Amazon Macie calculates this value based on the size of the latest version of each applicable object in those buckets. This value doesn't reflect the storage size of all versions of the applicable objects in the buckets.

lastUpdated :: Maybe POSIX

The date and time, in UTC and extended ISO 8601 format, when Amazon Macie most recently retrieved both bucket and object metadata from Amazon S3 for the buckets.

sizeInBytes :: Maybe Integer

The total storage size, in bytes, of the buckets.

If versioning is enabled for any of the buckets, Amazon Macie calculates this value based on the size of the latest version of each object in those buckets. This value doesn't reflect the storage size of all versions of the objects in the buckets.

bucketCountBySharedAccessType :: Maybe BucketCountBySharedAccessType

The total number of buckets that are or aren't shared with another Amazon Web Services account.

classifiableObjectCount :: Maybe Integer

The total number of objects that Amazon Macie can analyze in the buckets. These objects use a supported storage class and have a file name extension for a supported file or storage format.

unclassifiableObjectSizeInBytes :: Maybe ObjectLevelStatistics

The total storage size, in bytes, of the objects that Amazon Macie can't analyze in the buckets. These objects don't use a supported storage class or don't have a file name extension for a supported file or storage format.

unclassifiableObjectCount :: Maybe ObjectLevelStatistics

The total number of objects that Amazon Macie can't analyze in the buckets. These objects don't use a supported storage class or don't have a file name extension for a supported file or storage format.

bucketCount :: Maybe Integer

The total number of buckets.

bucketCountByEffectivePermission :: Maybe BucketCountByEffectivePermission

The total number of buckets that are publicly accessible based on a combination of permissions settings for each bucket.

bucketCountByObjectEncryptionRequirement :: Maybe BucketCountPolicyAllowsUnencryptedObjectUploads

The total number of buckets whose bucket policies do or don't require server-side encryption of objects when objects are uploaded to the buckets.

objectCount :: Maybe Integer

The total number of objects in the buckets.

classifiableSizeInBytes :: Maybe Integer

The total storage size, in bytes, of all the objects that Amazon Macie can analyze in the buckets. These objects use a supported storage class and have a file name extension for a supported file or storage format.

If versioning is enabled for any of the buckets, Macie calculates this value based on the size of the latest version of each applicable object in those buckets. This value doesn't reflect the storage size of all versions of all applicable objects in the buckets.

bucketCountByEncryptionType :: Maybe BucketCountByEncryptionType

The total number of buckets that use certain types of server-side encryption to encrypt new objects by default. This object also reports the total number of buckets that don't encrypt new objects by default.

httpStatus :: Int

The response's http status code.

newGetBucketStatisticsResponse

Arguments

:: Int
$sel:httpStatus:GetBucketStatisticsResponse'

-> GetBucketStatisticsResponse

Create a value of GetBucketStatisticsResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:sizeInBytesCompressed:GetBucketStatisticsResponse', getBucketStatisticsResponse_sizeInBytesCompressed - The total storage size, in bytes, of the objects that are compressed (.gz, .gzip, .zip) files in the buckets.

If versioning is enabled for any of the buckets, Amazon Macie calculates this value based on the size of the latest version of each applicable object in those buckets. This value doesn't reflect the storage size of all versions of the applicable objects in the buckets.

$sel:lastUpdated:GetBucketStatisticsResponse', getBucketStatisticsResponse_lastUpdated - The date and time, in UTC and extended ISO 8601 format, when Amazon Macie most recently retrieved both bucket and object metadata from Amazon S3 for the buckets.

$sel:sizeInBytes:GetBucketStatisticsResponse', getBucketStatisticsResponse_sizeInBytes - The total storage size, in bytes, of the buckets.

If versioning is enabled for any of the buckets, Amazon Macie calculates this value based on the size of the latest version of each object in those buckets. This value doesn't reflect the storage size of all versions of the objects in the buckets.

$sel:bucketCountBySharedAccessType:GetBucketStatisticsResponse', getBucketStatisticsResponse_bucketCountBySharedAccessType - The total number of buckets that are or aren't shared with another Amazon Web Services account.

$sel:classifiableObjectCount:GetBucketStatisticsResponse', getBucketStatisticsResponse_classifiableObjectCount - The total number of objects that Amazon Macie can analyze in the buckets. These objects use a supported storage class and have a file name extension for a supported file or storage format.

$sel:unclassifiableObjectSizeInBytes:GetBucketStatisticsResponse', getBucketStatisticsResponse_unclassifiableObjectSizeInBytes - The total storage size, in bytes, of the objects that Amazon Macie can't analyze in the buckets. These objects don't use a supported storage class or don't have a file name extension for a supported file or storage format.

$sel:unclassifiableObjectCount:GetBucketStatisticsResponse', getBucketStatisticsResponse_unclassifiableObjectCount - The total number of objects that Amazon Macie can't analyze in the buckets. These objects don't use a supported storage class or don't have a file name extension for a supported file or storage format.

$sel:bucketCount:GetBucketStatisticsResponse', getBucketStatisticsResponse_bucketCount - The total number of buckets.

$sel:bucketCountByEffectivePermission:GetBucketStatisticsResponse', getBucketStatisticsResponse_bucketCountByEffectivePermission - The total number of buckets that are publicly accessible based on a combination of permissions settings for each bucket.

$sel:bucketCountByObjectEncryptionRequirement:GetBucketStatisticsResponse', getBucketStatisticsResponse_bucketCountByObjectEncryptionRequirement - The total number of buckets whose bucket policies do or don't require server-side encryption of objects when objects are uploaded to the buckets.

$sel:objectCount:GetBucketStatisticsResponse', getBucketStatisticsResponse_objectCount - The total number of objects in the buckets.

$sel:classifiableSizeInBytes:GetBucketStatisticsResponse', getBucketStatisticsResponse_classifiableSizeInBytes - The total storage size, in bytes, of all the objects that Amazon Macie can analyze in the buckets. These objects use a supported storage class and have a file name extension for a supported file or storage format.

If versioning is enabled for any of the buckets, Macie calculates this value based on the size of the latest version of each applicable object in those buckets. This value doesn't reflect the storage size of all versions of all applicable objects in the buckets.

$sel:bucketCountByEncryptionType:GetBucketStatisticsResponse', getBucketStatisticsResponse_bucketCountByEncryptionType - The total number of buckets that use certain types of server-side encryption to encrypt new objects by default. This object also reports the total number of buckets that don't encrypt new objects by default.

$sel:httpStatus:GetBucketStatisticsResponse', getBucketStatisticsResponse_httpStatus - The response's http status code.

Response Lenses

getBucketStatisticsResponse_sizeInBytesCompressed :: Lens' GetBucketStatisticsResponse (Maybe Integer)

The total storage size, in bytes, of the objects that are compressed (.gz, .gzip, .zip) files in the buckets.

If versioning is enabled for any of the buckets, Amazon Macie calculates this value based on the size of the latest version of each applicable object in those buckets. This value doesn't reflect the storage size of all versions of the applicable objects in the buckets.

getBucketStatisticsResponse_lastUpdated :: Lens' GetBucketStatisticsResponse (Maybe UTCTime)

The date and time, in UTC and extended ISO 8601 format, when Amazon Macie most recently retrieved both bucket and object metadata from Amazon S3 for the buckets.

getBucketStatisticsResponse_sizeInBytes :: Lens' GetBucketStatisticsResponse (Maybe Integer)

The total storage size, in bytes, of the buckets.

If versioning is enabled for any of the buckets, Amazon Macie calculates this value based on the size of the latest version of each object in those buckets. This value doesn't reflect the storage size of all versions of the objects in the buckets.

getBucketStatisticsResponse_bucketCountBySharedAccessType :: Lens' GetBucketStatisticsResponse (Maybe BucketCountBySharedAccessType)

The total number of buckets that are or aren't shared with another Amazon Web Services account.

getBucketStatisticsResponse_classifiableObjectCount :: Lens' GetBucketStatisticsResponse (Maybe Integer)

The total number of objects that Amazon Macie can analyze in the buckets. These objects use a supported storage class and have a file name extension for a supported file or storage format.

getBucketStatisticsResponse_unclassifiableObjectSizeInBytes :: Lens' GetBucketStatisticsResponse (Maybe ObjectLevelStatistics)

The total storage size, in bytes, of the objects that Amazon Macie can't analyze in the buckets. These objects don't use a supported storage class or don't have a file name extension for a supported file or storage format.

getBucketStatisticsResponse_unclassifiableObjectCount :: Lens' GetBucketStatisticsResponse (Maybe ObjectLevelStatistics)

The total number of objects that Amazon Macie can't analyze in the buckets. These objects don't use a supported storage class or don't have a file name extension for a supported file or storage format.

getBucketStatisticsResponse_bucketCount :: Lens' GetBucketStatisticsResponse (Maybe Integer)

The total number of buckets.

getBucketStatisticsResponse_bucketCountByEffectivePermission :: Lens' GetBucketStatisticsResponse (Maybe BucketCountByEffectivePermission)

The total number of buckets that are publicly accessible based on a combination of permissions settings for each bucket.

getBucketStatisticsResponse_bucketCountByObjectEncryptionRequirement :: Lens' GetBucketStatisticsResponse (Maybe BucketCountPolicyAllowsUnencryptedObjectUploads)

The total number of buckets whose bucket policies do or don't require server-side encryption of objects when objects are uploaded to the buckets.

getBucketStatisticsResponse_objectCount :: Lens' GetBucketStatisticsResponse (Maybe Integer)

The total number of objects in the buckets.

getBucketStatisticsResponse_classifiableSizeInBytes :: Lens' GetBucketStatisticsResponse (Maybe Integer)

The total storage size, in bytes, of all the objects that Amazon Macie can analyze in the buckets. These objects use a supported storage class and have a file name extension for a supported file or storage format.

If versioning is enabled for any of the buckets, Macie calculates this value based on the size of the latest version of each applicable object in those buckets. This value doesn't reflect the storage size of all versions of all applicable objects in the buckets.

getBucketStatisticsResponse_bucketCountByEncryptionType :: Lens' GetBucketStatisticsResponse (Maybe BucketCountByEncryptionType)

The total number of buckets that use certain types of server-side encryption to encrypt new objects by default. This object also reports the total number of buckets that don't encrypt new objects by default.

getBucketStatisticsResponse_httpStatus :: Lens' GetBucketStatisticsResponse Int

The response's http status code.