Amazonka.MacieV2.GetClassificationExportConfiguration

Description

Retrieves the configuration settings for storing data classification results.

Creating a Request

data GetClassificationExportConfiguration

See: newGetClassificationExportConfiguration smart constructor.

Constructors

GetClassificationExportConfiguration'

newGetClassificationExportConfiguration :: GetClassificationExportConfiguration

Create a value of GetClassificationExportConfiguration with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

Destructuring the Response

data GetClassificationExportConfigurationResponse

See: newGetClassificationExportConfigurationResponse smart constructor.

Constructors

GetClassificationExportConfigurationResponse'

Fields

configuration :: Maybe ClassificationExportConfiguration

The location where data classification results are stored, and the encryption settings that are used when storing results in that location.

httpStatus :: Int

The response's http status code.

newGetClassificationExportConfigurationResponse

Arguments

:: Int
$sel:httpStatus:GetClassificationExportConfigurationResponse'

-> GetClassificationExportConfigurationResponse

Create a value of GetClassificationExportConfigurationResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:configuration:GetClassificationExportConfigurationResponse', getClassificationExportConfigurationResponse_configuration - The location where data classification results are stored, and the encryption settings that are used when storing results in that location.

$sel:httpStatus:GetClassificationExportConfigurationResponse', getClassificationExportConfigurationResponse_httpStatus - The response's http status code.

Response Lenses

getClassificationExportConfigurationResponse_configuration :: Lens' GetClassificationExportConfigurationResponse (Maybe ClassificationExportConfiguration)

The location where data classification results are stored, and the encryption settings that are used when storing results in that location.

getClassificationExportConfigurationResponse_httpStatus :: Lens' GetClassificationExportConfigurationResponse Int

The response's http status code.