Amazonka.MacieV2.GetUsageStatistics

Description

Retrieves (queries) quotas and aggregated usage data for one or more accounts.

This operation returns paginated results.

Creating a Request

data GetUsageStatistics

See: newGetUsageStatistics smart constructor.

Constructors

GetUsageStatistics'

Fields

timeRange :: Maybe TimeRange

The inclusive time period to query usage data for. Valid values are: MONTH_TO_DATE, for the current calendar month to date; and, PAST_30_DAYS, for the preceding 30 days. If you don't specify a value, Amazon Macie provides usage data for the preceding 30 days.

nextToken :: Maybe Text

The nextToken string that specifies which page of results to return in a paginated response.

filterBy :: Maybe [UsageStatisticsFilter]

An array of objects, one for each condition to use to filter the query results. If you specify more than one condition, Amazon Macie uses an AND operator to join the conditions.

maxResults :: Maybe Int

The maximum number of items to include in each page of the response.

sortBy :: Maybe UsageStatisticsSortBy

The criteria to use to sort the query results.

newGetUsageStatistics :: GetUsageStatistics

Create a value of GetUsageStatistics with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:timeRange:GetUsageStatistics', getUsageStatistics_timeRange - The inclusive time period to query usage data for. Valid values are: MONTH_TO_DATE, for the current calendar month to date; and, PAST_30_DAYS, for the preceding 30 days. If you don't specify a value, Amazon Macie provides usage data for the preceding 30 days.

$sel:nextToken:GetUsageStatistics', getUsageStatistics_nextToken - The nextToken string that specifies which page of results to return in a paginated response.

$sel:filterBy:GetUsageStatistics', getUsageStatistics_filterBy - An array of objects, one for each condition to use to filter the query results. If you specify more than one condition, Amazon Macie uses an AND operator to join the conditions.

$sel:maxResults:GetUsageStatistics', getUsageStatistics_maxResults - The maximum number of items to include in each page of the response.

$sel:sortBy:GetUsageStatistics', getUsageStatistics_sortBy - The criteria to use to sort the query results.

Request Lenses

getUsageStatistics_timeRange :: Lens' GetUsageStatistics (Maybe TimeRange)

The inclusive time period to query usage data for. Valid values are: MONTH_TO_DATE, for the current calendar month to date; and, PAST_30_DAYS, for the preceding 30 days. If you don't specify a value, Amazon Macie provides usage data for the preceding 30 days.

getUsageStatistics_nextToken :: Lens' GetUsageStatistics (Maybe Text)

The nextToken string that specifies which page of results to return in a paginated response.

getUsageStatistics_filterBy :: Lens' GetUsageStatistics (Maybe [UsageStatisticsFilter])

An array of objects, one for each condition to use to filter the query results. If you specify more than one condition, Amazon Macie uses an AND operator to join the conditions.

getUsageStatistics_maxResults :: Lens' GetUsageStatistics (Maybe Int)

The maximum number of items to include in each page of the response.

getUsageStatistics_sortBy :: Lens' GetUsageStatistics (Maybe UsageStatisticsSortBy)

The criteria to use to sort the query results.

Destructuring the Response

data GetUsageStatisticsResponse

See: newGetUsageStatisticsResponse smart constructor.

Constructors

GetUsageStatisticsResponse'

Fields

timeRange :: Maybe TimeRange

The inclusive time period that the usage data applies to. Possible values are: MONTH_TO_DATE, for the current calendar month to date; and, PAST_30_DAYS, for the preceding 30 days.

records :: Maybe [UsageRecord]

An array of objects that contains the results of the query. Each object contains the data for an account that meets the filter criteria specified in the request.

nextToken :: Maybe Text

The string to use in a subsequent request to get the next page of results in a paginated response. This value is null if there are no additional pages.

httpStatus :: Int

The response's http status code.

newGetUsageStatisticsResponse

Arguments

:: Int
$sel:httpStatus:GetUsageStatisticsResponse'

-> GetUsageStatisticsResponse

Create a value of GetUsageStatisticsResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:timeRange:GetUsageStatistics', getUsageStatisticsResponse_timeRange - The inclusive time period that the usage data applies to. Possible values are: MONTH_TO_DATE, for the current calendar month to date; and, PAST_30_DAYS, for the preceding 30 days.

$sel:records:GetUsageStatisticsResponse', getUsageStatisticsResponse_records - An array of objects that contains the results of the query. Each object contains the data for an account that meets the filter criteria specified in the request.

$sel:nextToken:GetUsageStatistics', getUsageStatisticsResponse_nextToken - The string to use in a subsequent request to get the next page of results in a paginated response. This value is null if there are no additional pages.

$sel:httpStatus:GetUsageStatisticsResponse', getUsageStatisticsResponse_httpStatus - The response's http status code.

Response Lenses

getUsageStatisticsResponse_timeRange :: Lens' GetUsageStatisticsResponse (Maybe TimeRange)

The inclusive time period that the usage data applies to. Possible values are: MONTH_TO_DATE, for the current calendar month to date; and, PAST_30_DAYS, for the preceding 30 days.

getUsageStatisticsResponse_records :: Lens' GetUsageStatisticsResponse (Maybe [UsageRecord])

An array of objects that contains the results of the query. Each object contains the data for an account that meets the filter criteria specified in the request.

getUsageStatisticsResponse_nextToken :: Lens' GetUsageStatisticsResponse (Maybe Text)

The string to use in a subsequent request to get the next page of results in a paginated response. This value is null if there are no additional pages.

getUsageStatisticsResponse_httpStatus :: Lens' GetUsageStatisticsResponse Int

The response's http status code.