Amazonka.MacieV2.GetUsageTotals

Description

Retrieves (queries) aggregated usage data for an account.

Creating a Request

data GetUsageTotals

See: newGetUsageTotals smart constructor.

Constructors

GetUsageTotals'

Fields

timeRange :: Maybe Text

The inclusive time period to retrieve the data for. Valid values are: MONTH_TO_DATE, for the current calendar month to date; and, PAST_30_DAYS, for the preceding 30 days. If you don't specify a value for this parameter, Amazon Macie provides aggregated usage data for the preceding 30 days.

newGetUsageTotals :: GetUsageTotals

Create a value of GetUsageTotals with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:timeRange:GetUsageTotals', getUsageTotals_timeRange - The inclusive time period to retrieve the data for. Valid values are: MONTH_TO_DATE, for the current calendar month to date; and, PAST_30_DAYS, for the preceding 30 days. If you don't specify a value for this parameter, Amazon Macie provides aggregated usage data for the preceding 30 days.

Request Lenses

getUsageTotals_timeRange :: Lens' GetUsageTotals (Maybe Text)

The inclusive time period to retrieve the data for. Valid values are: MONTH_TO_DATE, for the current calendar month to date; and, PAST_30_DAYS, for the preceding 30 days. If you don't specify a value for this parameter, Amazon Macie provides aggregated usage data for the preceding 30 days.

Destructuring the Response

data GetUsageTotalsResponse

See: newGetUsageTotalsResponse smart constructor.

Constructors

GetUsageTotalsResponse'

Fields

timeRange :: Maybe TimeRange

The inclusive time period that the usage data applies to. Possible values are: MONTH_TO_DATE, for the current calendar month to date; and, PAST_30_DAYS, for the preceding 30 days.

usageTotals :: Maybe [UsageTotal]

An array of objects that contains the results of the query. Each object contains the data for a specific usage metric.

httpStatus :: Int

The response's http status code.

newGetUsageTotalsResponse

Arguments

:: Int
$sel:httpStatus:GetUsageTotalsResponse'

-> GetUsageTotalsResponse

Create a value of GetUsageTotalsResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:timeRange:GetUsageTotals', getUsageTotalsResponse_timeRange - The inclusive time period that the usage data applies to. Possible values are: MONTH_TO_DATE, for the current calendar month to date; and, PAST_30_DAYS, for the preceding 30 days.

$sel:usageTotals:GetUsageTotalsResponse', getUsageTotalsResponse_usageTotals - An array of objects that contains the results of the query. Each object contains the data for a specific usage metric.

$sel:httpStatus:GetUsageTotalsResponse', getUsageTotalsResponse_httpStatus - The response's http status code.

Response Lenses

getUsageTotalsResponse_timeRange :: Lens' GetUsageTotalsResponse (Maybe TimeRange)

The inclusive time period that the usage data applies to. Possible values are: MONTH_TO_DATE, for the current calendar month to date; and, PAST_30_DAYS, for the preceding 30 days.

getUsageTotalsResponse_usageTotals :: Lens' GetUsageTotalsResponse (Maybe [UsageTotal])

An array of objects that contains the results of the query. Each object contains the data for a specific usage metric.

getUsageTotalsResponse_httpStatus :: Lens' GetUsageTotalsResponse Int

The response's http status code.