Amazonka.MacieV2.ListClassificationJobs

Description

Retrieves a subset of information about one or more classification jobs.

This operation returns paginated results.

Creating a Request

data ListClassificationJobs

See: newListClassificationJobs smart constructor.

Constructors

ListClassificationJobs'

Fields

filterCriteria :: Maybe ListJobsFilterCriteria

The criteria to use to filter the results.

sortCriteria :: Maybe ListJobsSortCriteria

The criteria to use to sort the results.

nextToken :: Maybe Text

The nextToken string that specifies which page of results to return in a paginated response.

maxResults :: Maybe Int

The maximum number of items to include in each page of the response.

newListClassificationJobs :: ListClassificationJobs

Create a value of ListClassificationJobs with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:filterCriteria:ListClassificationJobs', listClassificationJobs_filterCriteria - The criteria to use to filter the results.

$sel:sortCriteria:ListClassificationJobs', listClassificationJobs_sortCriteria - The criteria to use to sort the results.

$sel:nextToken:ListClassificationJobs', listClassificationJobs_nextToken - The nextToken string that specifies which page of results to return in a paginated response.

$sel:maxResults:ListClassificationJobs', listClassificationJobs_maxResults - The maximum number of items to include in each page of the response.

Request Lenses

listClassificationJobs_filterCriteria :: Lens' ListClassificationJobs (Maybe ListJobsFilterCriteria)

The criteria to use to filter the results.

listClassificationJobs_sortCriteria :: Lens' ListClassificationJobs (Maybe ListJobsSortCriteria)

The criteria to use to sort the results.

listClassificationJobs_nextToken :: Lens' ListClassificationJobs (Maybe Text)

The nextToken string that specifies which page of results to return in a paginated response.

listClassificationJobs_maxResults :: Lens' ListClassificationJobs (Maybe Int)

The maximum number of items to include in each page of the response.

Destructuring the Response

data ListClassificationJobsResponse

See: newListClassificationJobsResponse smart constructor.

Constructors

ListClassificationJobsResponse'

Fields

items :: Maybe [JobSummary]

An array of objects, one for each job that meets the filter criteria specified in the request.

nextToken :: Maybe Text

The string to use in a subsequent request to get the next page of results in a paginated response. This value is null if there are no additional pages.

httpStatus :: Int

The response's http status code.

newListClassificationJobsResponse

Arguments

:: Int
$sel:httpStatus:ListClassificationJobsResponse'

-> ListClassificationJobsResponse

Create a value of ListClassificationJobsResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:items:ListClassificationJobsResponse', listClassificationJobsResponse_items - An array of objects, one for each job that meets the filter criteria specified in the request.

$sel:nextToken:ListClassificationJobs', listClassificationJobsResponse_nextToken - The string to use in a subsequent request to get the next page of results in a paginated response. This value is null if there are no additional pages.

$sel:httpStatus:ListClassificationJobsResponse', listClassificationJobsResponse_httpStatus - The response's http status code.

Response Lenses

listClassificationJobsResponse_items :: Lens' ListClassificationJobsResponse (Maybe [JobSummary])

An array of objects, one for each job that meets the filter criteria specified in the request.

listClassificationJobsResponse_nextToken :: Lens' ListClassificationJobsResponse (Maybe Text)

The string to use in a subsequent request to get the next page of results in a paginated response. This value is null if there are no additional pages.

listClassificationJobsResponse_httpStatus :: Lens' ListClassificationJobsResponse Int

The response's http status code.