Amazonka.MacieV2.ListManagedDataIdentifiers

Description

Retrieves information about all the managed data identifiers that Amazon Macie currently provides.

Creating a Request

data ListManagedDataIdentifiers

See: newListManagedDataIdentifiers smart constructor.

Constructors

ListManagedDataIdentifiers'

Fields

nextToken :: Maybe Text

The nextToken string that specifies which page of results to return in a paginated response.

newListManagedDataIdentifiers :: ListManagedDataIdentifiers

Create a value of ListManagedDataIdentifiers with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:nextToken:ListManagedDataIdentifiers', listManagedDataIdentifiers_nextToken - The nextToken string that specifies which page of results to return in a paginated response.

Request Lenses

listManagedDataIdentifiers_nextToken :: Lens' ListManagedDataIdentifiers (Maybe Text)

The nextToken string that specifies which page of results to return in a paginated response.

Destructuring the Response

data ListManagedDataIdentifiersResponse

See: newListManagedDataIdentifiersResponse smart constructor.

Constructors

ListManagedDataIdentifiersResponse'

Fields

items :: Maybe [ManagedDataIdentifierSummary]

An array of objects, one for each managed data identifier.

nextToken :: Maybe Text

The string to use in a subsequent request to get the next page of results in a paginated response. This value is null if there are no additional pages.

httpStatus :: Int

The response's http status code.

newListManagedDataIdentifiersResponse

Arguments

:: Int
$sel:httpStatus:ListManagedDataIdentifiersResponse'

-> ListManagedDataIdentifiersResponse

Create a value of ListManagedDataIdentifiersResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:items:ListManagedDataIdentifiersResponse', listManagedDataIdentifiersResponse_items - An array of objects, one for each managed data identifier.

$sel:nextToken:ListManagedDataIdentifiers', listManagedDataIdentifiersResponse_nextToken - The string to use in a subsequent request to get the next page of results in a paginated response. This value is null if there are no additional pages.

$sel:httpStatus:ListManagedDataIdentifiersResponse', listManagedDataIdentifiersResponse_httpStatus - The response's http status code.

Response Lenses

listManagedDataIdentifiersResponse_items :: Lens' ListManagedDataIdentifiersResponse (Maybe [ManagedDataIdentifierSummary])

An array of objects, one for each managed data identifier.

listManagedDataIdentifiersResponse_nextToken :: Lens' ListManagedDataIdentifiersResponse (Maybe Text)

The string to use in a subsequent request to get the next page of results in a paginated response. This value is null if there are no additional pages.

listManagedDataIdentifiersResponse_httpStatus :: Lens' ListManagedDataIdentifiersResponse Int

The response's http status code.