Amazonka.MacieV2.ListOrganizationAdminAccounts

Description

Retrieves information about the delegated Amazon Macie administrator account for an Amazon Web Services organization.

This operation returns paginated results.

Creating a Request

data ListOrganizationAdminAccounts

See: newListOrganizationAdminAccounts smart constructor.

Constructors

ListOrganizationAdminAccounts'

Fields

nextToken :: Maybe Text

The nextToken string that specifies which page of results to return in a paginated response.

maxResults :: Maybe Natural

The maximum number of items to include in each page of a paginated response.

newListOrganizationAdminAccounts :: ListOrganizationAdminAccounts

Create a value of ListOrganizationAdminAccounts with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:nextToken:ListOrganizationAdminAccounts', listOrganizationAdminAccounts_nextToken - The nextToken string that specifies which page of results to return in a paginated response.

$sel:maxResults:ListOrganizationAdminAccounts', listOrganizationAdminAccounts_maxResults - The maximum number of items to include in each page of a paginated response.

Request Lenses

listOrganizationAdminAccounts_nextToken :: Lens' ListOrganizationAdminAccounts (Maybe Text)

The nextToken string that specifies which page of results to return in a paginated response.

listOrganizationAdminAccounts_maxResults :: Lens' ListOrganizationAdminAccounts (Maybe Natural)

The maximum number of items to include in each page of a paginated response.

Destructuring the Response

data ListOrganizationAdminAccountsResponse

See: newListOrganizationAdminAccountsResponse smart constructor.

Constructors

ListOrganizationAdminAccountsResponse'

Fields

adminAccounts :: Maybe [AdminAccount]

An array of objects, one for each delegated Amazon Macie administrator account for the organization. Only one of these accounts can have a status of ENABLED.

nextToken :: Maybe Text

The string to use in a subsequent request to get the next page of results in a paginated response. This value is null if there are no additional pages.

httpStatus :: Int

The response's http status code.

newListOrganizationAdminAccountsResponse

Arguments

:: Int
$sel:httpStatus:ListOrganizationAdminAccountsResponse'

-> ListOrganizationAdminAccountsResponse

Create a value of ListOrganizationAdminAccountsResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:adminAccounts:ListOrganizationAdminAccountsResponse', listOrganizationAdminAccountsResponse_adminAccounts - An array of objects, one for each delegated Amazon Macie administrator account for the organization. Only one of these accounts can have a status of ENABLED.

$sel:nextToken:ListOrganizationAdminAccounts', listOrganizationAdminAccountsResponse_nextToken - The string to use in a subsequent request to get the next page of results in a paginated response. This value is null if there are no additional pages.

$sel:httpStatus:ListOrganizationAdminAccountsResponse', listOrganizationAdminAccountsResponse_httpStatus - The response's http status code.

Response Lenses

listOrganizationAdminAccountsResponse_adminAccounts :: Lens' ListOrganizationAdminAccountsResponse (Maybe [AdminAccount])

An array of objects, one for each delegated Amazon Macie administrator account for the organization. Only one of these accounts can have a status of ENABLED.

listOrganizationAdminAccountsResponse_nextToken :: Lens' ListOrganizationAdminAccountsResponse (Maybe Text)

The string to use in a subsequent request to get the next page of results in a paginated response. This value is null if there are no additional pages.

listOrganizationAdminAccountsResponse_httpStatus :: Lens' ListOrganizationAdminAccountsResponse Int

The response's http status code.