Amazonka.MacieV2.Types.AssumedRole

Description

Documentation

data AssumedRole

Provides information about an identity that performed an action on an affected resource by using temporary security credentials. The credentials were obtained using the AssumeRole operation of the Security Token Service (STS) API.

See: newAssumedRole smart constructor.

Constructors

AssumedRole'

Fields

principalId :: Maybe Text

The unique identifier for the entity that was used to get the credentials.

arn :: Maybe Text

The Amazon Resource Name (ARN) of the entity that was used to get the credentials.

sessionContext :: Maybe SessionContext

The details of the session that was created for the credentials, including the entity that issued the session.

accountId :: Maybe Text

The unique identifier for the Amazon Web Services account that owns the entity that was used to get the credentials.

accessKeyId :: Maybe Text

The Amazon Web Services access key ID that identifies the credentials.

newAssumedRole :: AssumedRole

Create a value of AssumedRole with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:principalId:AssumedRole', assumedRole_principalId - The unique identifier for the entity that was used to get the credentials.

$sel:arn:AssumedRole', assumedRole_arn - The Amazon Resource Name (ARN) of the entity that was used to get the credentials.

$sel:sessionContext:AssumedRole', assumedRole_sessionContext - The details of the session that was created for the credentials, including the entity that issued the session.

$sel:accountId:AssumedRole', assumedRole_accountId - The unique identifier for the Amazon Web Services account that owns the entity that was used to get the credentials.

$sel:accessKeyId:AssumedRole', assumedRole_accessKeyId - The Amazon Web Services access key ID that identifies the credentials.

assumedRole_principalId :: Lens' AssumedRole (Maybe Text)

The unique identifier for the entity that was used to get the credentials.

assumedRole_arn :: Lens' AssumedRole (Maybe Text)

The Amazon Resource Name (ARN) of the entity that was used to get the credentials.

assumedRole_sessionContext :: Lens' AssumedRole (Maybe SessionContext)

The details of the session that was created for the credentials, including the entity that issued the session.

assumedRole_accountId :: Lens' AssumedRole (Maybe Text)

The unique identifier for the Amazon Web Services account that owns the entity that was used to get the credentials.

assumedRole_accessKeyId :: Lens' AssumedRole (Maybe Text)

The Amazon Web Services access key ID that identifies the credentials.