Amazonka.MacieV2.Types.BucketCountBySharedAccessType

Description

Documentation

data BucketCountBySharedAccessType

Provides information about the number of S3 buckets that are or aren't shared with other Amazon Web Services accounts.

See: newBucketCountBySharedAccessType smart constructor.

Constructors

BucketCountBySharedAccessType'

Fields

notShared :: Maybe Integer

The total number of buckets that aren't shared with other Amazon Web Services accounts.

internal :: Maybe Integer

The total number of buckets that are shared with an Amazon Web Services account that's part of the same Amazon Macie organization.

external :: Maybe Integer

The total number of buckets that are shared with an Amazon Web Services account that isn't part of the same Amazon Macie organization.

unknown :: Maybe Integer

The total number of buckets that Amazon Macie wasn't able to evaluate shared access settings for. Macie can't determine whether these buckets are shared with other Amazon Web Services accounts.

newBucketCountBySharedAccessType :: BucketCountBySharedAccessType

Create a value of BucketCountBySharedAccessType with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:notShared:BucketCountBySharedAccessType', bucketCountBySharedAccessType_notShared - The total number of buckets that aren't shared with other Amazon Web Services accounts.

$sel:internal:BucketCountBySharedAccessType', bucketCountBySharedAccessType_internal - The total number of buckets that are shared with an Amazon Web Services account that's part of the same Amazon Macie organization.

$sel:external:BucketCountBySharedAccessType', bucketCountBySharedAccessType_external - The total number of buckets that are shared with an Amazon Web Services account that isn't part of the same Amazon Macie organization.

$sel:unknown:BucketCountBySharedAccessType', bucketCountBySharedAccessType_unknown - The total number of buckets that Amazon Macie wasn't able to evaluate shared access settings for. Macie can't determine whether these buckets are shared with other Amazon Web Services accounts.

bucketCountBySharedAccessType_notShared :: Lens' BucketCountBySharedAccessType (Maybe Integer)

The total number of buckets that aren't shared with other Amazon Web Services accounts.

bucketCountBySharedAccessType_internal :: Lens' BucketCountBySharedAccessType (Maybe Integer)

The total number of buckets that are shared with an Amazon Web Services account that's part of the same Amazon Macie organization.

bucketCountBySharedAccessType_external :: Lens' BucketCountBySharedAccessType (Maybe Integer)

The total number of buckets that are shared with an Amazon Web Services account that isn't part of the same Amazon Macie organization.

bucketCountBySharedAccessType_unknown :: Lens' BucketCountBySharedAccessType (Maybe Integer)

The total number of buckets that Amazon Macie wasn't able to evaluate shared access settings for. Macie can't determine whether these buckets are shared with other Amazon Web Services accounts.