Amazonka.MacieV2.Types.BucketMetadata

Description

Documentation

data BucketMetadata

Provides statistical data and other information about an S3 bucket that Amazon Macie monitors and analyzes for your account. If an error occurs when Macie attempts to retrieve and process information about the bucket or the bucket's objects, the value for most of these properties is null. Exceptions are accountId, bucketArn, bucketCreatedAt, bucketName, lastUpdated, and region. To identify the cause of the error, refer to the errorCode and errorMessage values.

See: newBucketMetadata smart constructor.

Constructors

BucketMetadata'

Fields

sizeInBytesCompressed :: Maybe Integer

The total storage size, in bytes, of the objects that are compressed (.gz, .gzip, .zip) files in the bucket.

If versioning is enabled for the bucket, Amazon Macie calculates this value based on the size of the latest version of each applicable object in the bucket. This value doesn't reflect the storage size of all versions of each applicable object in the bucket.

lastUpdated :: Maybe POSIX

The date and time, in UTC and extended ISO 8601 format, when Amazon Macie most recently retrieved both bucket and object metadata from Amazon S3 for the bucket.

objectCountByEncryptionType :: Maybe ObjectCountByEncryptionType

The total number of objects that are in the bucket, grouped by server-side encryption type. This includes a grouping that reports the total number of objects that aren't encrypted or use client-side encryption.

sharedAccess :: Maybe SharedAccess

Specifies whether the bucket is shared with another Amazon Web Services account. Possible values are:

EXTERNAL - The bucket is shared with an Amazon Web Services account that isn't part of the same Amazon Macie organization.

INTERNAL - The bucket is shared with an Amazon Web Services account that's part of the same Amazon Macie organization.

NOT_SHARED - The bucket isn't shared with other Amazon Web Services accounts.

UNKNOWN - Amazon Macie wasn't able to evaluate the shared access settings for the bucket.

sizeInBytes :: Maybe Integer

The total storage size, in bytes, of the bucket.

If versioning is enabled for the bucket, Amazon Macie calculates this value based on the size of the latest version of each object in the bucket. This value doesn't reflect the storage size of all versions of each object in the bucket.

bucketName :: Maybe Text

The name of the bucket.

classifiableObjectCount :: Maybe Integer

The total number of objects that Amazon Macie can analyze in the bucket. These objects use a supported storage class and have a file name extension for a supported file or storage format.

accountId :: Maybe Text

The unique identifier for the Amazon Web Services account that owns the bucket.

unclassifiableObjectSizeInBytes :: Maybe ObjectLevelStatistics

The total storage size, in bytes, of the objects that Amazon Macie can't analyze in the bucket. These objects don't use a supported storage class or don't have a file name extension for a supported file or storage format.

unclassifiableObjectCount :: Maybe ObjectLevelStatistics

The total number of objects that Amazon Macie can't analyze in the bucket. These objects don't use a supported storage class or don't have a file name extension for a supported file or storage format.

versioning :: Maybe Bool

Specifies whether versioning is enabled for the bucket.

allowsUnencryptedObjectUploads :: Maybe AllowsUnencryptedObjectUploads

Specifies whether the bucket policy for the bucket requires server-side encryption of objects when objects are uploaded to the bucket. Possible values are:

FALSE - The bucket policy requires server-side encryption of new objects. PutObject requests must include the x-amz-server-side-encryption header and the value for that header must be AES256 or aws:kms.

TRUE - The bucket doesn't have a bucket policy or it has a bucket policy that doesn't require server-side encryption of new objects. If a bucket policy exists, it doesn't require PutObject requests to include the x-amz-server-side-encryption header and it doesn't require the value for that header to be AES256 or aws:kms.

UNKNOWN - Amazon Macie can't determine whether the bucket policy requires server-side encryption of new objects.

publicAccess :: Maybe BucketPublicAccess

Specifies whether the bucket is publicly accessible due to the combination of permissions settings that apply to the bucket, and provides information about those settings.

errorCode :: Maybe BucketMetadataErrorCode

Specifies the error code for an error that prevented Amazon Macie from retrieving and processing information about the bucket and the bucket's objects. If this value is ACCESS_DENIED, Macie doesn't have permission to retrieve the information. For example, the bucket has a restrictive bucket policy and Amazon S3 denied the request. If this value is null, Macie was able to retrieve and process the information.

jobDetails :: Maybe JobDetails

Specifies whether any one-time or recurring classification jobs are configured to analyze data in the bucket, and, if so, the details of the job that ran most recently.

region :: Maybe Text

The Amazon Web Services Region that hosts the bucket.

bucketCreatedAt :: Maybe POSIX

The date and time, in UTC and extended ISO 8601 format, when the bucket was created.

errorMessage :: Maybe Text

A brief description of the error (errorCode) that prevented Amazon Macie from retrieving and processing information about the bucket and the bucket's objects. This value is null if Macie was able to retrieve and process the information.

bucketArn :: Maybe Text

The Amazon Resource Name (ARN) of the bucket.

objectCount :: Maybe Integer

The total number of objects in the bucket.

replicationDetails :: Maybe ReplicationDetails

Specifies whether the bucket is configured to replicate one or more objects to buckets for other Amazon Web Services accounts and, if so, which accounts.

serverSideEncryption :: Maybe BucketServerSideEncryption

Specifies whether the bucket encrypts new objects by default and, if so, the type of server-side encryption that's used.

tags :: Maybe [KeyValuePair]

An array that specifies the tags (keys and values) that are associated with the bucket.

classifiableSizeInBytes :: Maybe Integer

The total storage size, in bytes, of the objects that Amazon Macie can analyze in the bucket. These objects use a supported storage class and have a file name extension for a supported file or storage format.

If versioning is enabled for the bucket, Macie calculates this value based on the size of the latest version of each applicable object in the bucket. This value doesn't reflect the storage size of all versions of each applicable object in the bucket.

newBucketMetadata :: BucketMetadata

Create a value of BucketMetadata with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:sizeInBytesCompressed:BucketMetadata', bucketMetadata_sizeInBytesCompressed - The total storage size, in bytes, of the objects that are compressed (.gz, .gzip, .zip) files in the bucket.

If versioning is enabled for the bucket, Amazon Macie calculates this value based on the size of the latest version of each applicable object in the bucket. This value doesn't reflect the storage size of all versions of each applicable object in the bucket.

$sel:lastUpdated:BucketMetadata', bucketMetadata_lastUpdated - The date and time, in UTC and extended ISO 8601 format, when Amazon Macie most recently retrieved both bucket and object metadata from Amazon S3 for the bucket.

$sel:objectCountByEncryptionType:BucketMetadata', bucketMetadata_objectCountByEncryptionType - The total number of objects that are in the bucket, grouped by server-side encryption type. This includes a grouping that reports the total number of objects that aren't encrypted or use client-side encryption.

$sel:sharedAccess:BucketMetadata', bucketMetadata_sharedAccess - Specifies whether the bucket is shared with another Amazon Web Services account. Possible values are:

EXTERNAL - The bucket is shared with an Amazon Web Services account that isn't part of the same Amazon Macie organization.

INTERNAL - The bucket is shared with an Amazon Web Services account that's part of the same Amazon Macie organization.

NOT_SHARED - The bucket isn't shared with other Amazon Web Services accounts.

UNKNOWN - Amazon Macie wasn't able to evaluate the shared access settings for the bucket.

$sel:sizeInBytes:BucketMetadata', bucketMetadata_sizeInBytes - The total storage size, in bytes, of the bucket.

If versioning is enabled for the bucket, Amazon Macie calculates this value based on the size of the latest version of each object in the bucket. This value doesn't reflect the storage size of all versions of each object in the bucket.

$sel:bucketName:BucketMetadata', bucketMetadata_bucketName - The name of the bucket.

$sel:classifiableObjectCount:BucketMetadata', bucketMetadata_classifiableObjectCount - The total number of objects that Amazon Macie can analyze in the bucket. These objects use a supported storage class and have a file name extension for a supported file or storage format.

$sel:accountId:BucketMetadata', bucketMetadata_accountId - The unique identifier for the Amazon Web Services account that owns the bucket.

$sel:unclassifiableObjectSizeInBytes:BucketMetadata', bucketMetadata_unclassifiableObjectSizeInBytes - The total storage size, in bytes, of the objects that Amazon Macie can't analyze in the bucket. These objects don't use a supported storage class or don't have a file name extension for a supported file or storage format.

$sel:unclassifiableObjectCount:BucketMetadata', bucketMetadata_unclassifiableObjectCount - The total number of objects that Amazon Macie can't analyze in the bucket. These objects don't use a supported storage class or don't have a file name extension for a supported file or storage format.

$sel:versioning:BucketMetadata', bucketMetadata_versioning - Specifies whether versioning is enabled for the bucket.

$sel:allowsUnencryptedObjectUploads:BucketMetadata', bucketMetadata_allowsUnencryptedObjectUploads - Specifies whether the bucket policy for the bucket requires server-side encryption of objects when objects are uploaded to the bucket. Possible values are:

FALSE - The bucket policy requires server-side encryption of new objects. PutObject requests must include the x-amz-server-side-encryption header and the value for that header must be AES256 or aws:kms.

TRUE - The bucket doesn't have a bucket policy or it has a bucket policy that doesn't require server-side encryption of new objects. If a bucket policy exists, it doesn't require PutObject requests to include the x-amz-server-side-encryption header and it doesn't require the value for that header to be AES256 or aws:kms.

UNKNOWN - Amazon Macie can't determine whether the bucket policy requires server-side encryption of new objects.

$sel:publicAccess:BucketMetadata', bucketMetadata_publicAccess - Specifies whether the bucket is publicly accessible due to the combination of permissions settings that apply to the bucket, and provides information about those settings.

$sel:errorCode:BucketMetadata', bucketMetadata_errorCode - Specifies the error code for an error that prevented Amazon Macie from retrieving and processing information about the bucket and the bucket's objects. If this value is ACCESS_DENIED, Macie doesn't have permission to retrieve the information. For example, the bucket has a restrictive bucket policy and Amazon S3 denied the request. If this value is null, Macie was able to retrieve and process the information.

$sel:jobDetails:BucketMetadata', bucketMetadata_jobDetails - Specifies whether any one-time or recurring classification jobs are configured to analyze data in the bucket, and, if so, the details of the job that ran most recently.

$sel:region:BucketMetadata', bucketMetadata_region - The Amazon Web Services Region that hosts the bucket.

$sel:bucketCreatedAt:BucketMetadata', bucketMetadata_bucketCreatedAt - The date and time, in UTC and extended ISO 8601 format, when the bucket was created.

$sel:errorMessage:BucketMetadata', bucketMetadata_errorMessage - A brief description of the error (errorCode) that prevented Amazon Macie from retrieving and processing information about the bucket and the bucket's objects. This value is null if Macie was able to retrieve and process the information.

$sel:bucketArn:BucketMetadata', bucketMetadata_bucketArn - The Amazon Resource Name (ARN) of the bucket.

$sel:objectCount:BucketMetadata', bucketMetadata_objectCount - The total number of objects in the bucket.

$sel:replicationDetails:BucketMetadata', bucketMetadata_replicationDetails - Specifies whether the bucket is configured to replicate one or more objects to buckets for other Amazon Web Services accounts and, if so, which accounts.

$sel:serverSideEncryption:BucketMetadata', bucketMetadata_serverSideEncryption - Specifies whether the bucket encrypts new objects by default and, if so, the type of server-side encryption that's used.

$sel:tags:BucketMetadata', bucketMetadata_tags - An array that specifies the tags (keys and values) that are associated with the bucket.

$sel:classifiableSizeInBytes:BucketMetadata', bucketMetadata_classifiableSizeInBytes - The total storage size, in bytes, of the objects that Amazon Macie can analyze in the bucket. These objects use a supported storage class and have a file name extension for a supported file or storage format.

If versioning is enabled for the bucket, Macie calculates this value based on the size of the latest version of each applicable object in the bucket. This value doesn't reflect the storage size of all versions of each applicable object in the bucket.

bucketMetadata_sizeInBytesCompressed :: Lens' BucketMetadata (Maybe Integer)

The total storage size, in bytes, of the objects that are compressed (.gz, .gzip, .zip) files in the bucket.

If versioning is enabled for the bucket, Amazon Macie calculates this value based on the size of the latest version of each applicable object in the bucket. This value doesn't reflect the storage size of all versions of each applicable object in the bucket.

bucketMetadata_lastUpdated :: Lens' BucketMetadata (Maybe UTCTime)

The date and time, in UTC and extended ISO 8601 format, when Amazon Macie most recently retrieved both bucket and object metadata from Amazon S3 for the bucket.

bucketMetadata_objectCountByEncryptionType :: Lens' BucketMetadata (Maybe ObjectCountByEncryptionType)

The total number of objects that are in the bucket, grouped by server-side encryption type. This includes a grouping that reports the total number of objects that aren't encrypted or use client-side encryption.

bucketMetadata_sharedAccess :: Lens' BucketMetadata (Maybe SharedAccess)

Specifies whether the bucket is shared with another Amazon Web Services account. Possible values are:

EXTERNAL - The bucket is shared with an Amazon Web Services account that isn't part of the same Amazon Macie organization.

INTERNAL - The bucket is shared with an Amazon Web Services account that's part of the same Amazon Macie organization.

NOT_SHARED - The bucket isn't shared with other Amazon Web Services accounts.

UNKNOWN - Amazon Macie wasn't able to evaluate the shared access settings for the bucket.

bucketMetadata_sizeInBytes :: Lens' BucketMetadata (Maybe Integer)

The total storage size, in bytes, of the bucket.

If versioning is enabled for the bucket, Amazon Macie calculates this value based on the size of the latest version of each object in the bucket. This value doesn't reflect the storage size of all versions of each object in the bucket.

bucketMetadata_bucketName :: Lens' BucketMetadata (Maybe Text)

The name of the bucket.

bucketMetadata_classifiableObjectCount :: Lens' BucketMetadata (Maybe Integer)

The total number of objects that Amazon Macie can analyze in the bucket. These objects use a supported storage class and have a file name extension for a supported file or storage format.

bucketMetadata_accountId :: Lens' BucketMetadata (Maybe Text)

The unique identifier for the Amazon Web Services account that owns the bucket.

bucketMetadata_unclassifiableObjectSizeInBytes :: Lens' BucketMetadata (Maybe ObjectLevelStatistics)

The total storage size, in bytes, of the objects that Amazon Macie can't analyze in the bucket. These objects don't use a supported storage class or don't have a file name extension for a supported file or storage format.

bucketMetadata_unclassifiableObjectCount :: Lens' BucketMetadata (Maybe ObjectLevelStatistics)

The total number of objects that Amazon Macie can't analyze in the bucket. These objects don't use a supported storage class or don't have a file name extension for a supported file or storage format.

bucketMetadata_versioning :: Lens' BucketMetadata (Maybe Bool)

Specifies whether versioning is enabled for the bucket.

bucketMetadata_allowsUnencryptedObjectUploads :: Lens' BucketMetadata (Maybe AllowsUnencryptedObjectUploads)

Specifies whether the bucket policy for the bucket requires server-side encryption of objects when objects are uploaded to the bucket. Possible values are:

FALSE - The bucket policy requires server-side encryption of new objects. PutObject requests must include the x-amz-server-side-encryption header and the value for that header must be AES256 or aws:kms.

TRUE - The bucket doesn't have a bucket policy or it has a bucket policy that doesn't require server-side encryption of new objects. If a bucket policy exists, it doesn't require PutObject requests to include the x-amz-server-side-encryption header and it doesn't require the value for that header to be AES256 or aws:kms.

UNKNOWN - Amazon Macie can't determine whether the bucket policy requires server-side encryption of new objects.

bucketMetadata_publicAccess :: Lens' BucketMetadata (Maybe BucketPublicAccess)

Specifies whether the bucket is publicly accessible due to the combination of permissions settings that apply to the bucket, and provides information about those settings.

bucketMetadata_errorCode :: Lens' BucketMetadata (Maybe BucketMetadataErrorCode)

Specifies the error code for an error that prevented Amazon Macie from retrieving and processing information about the bucket and the bucket's objects. If this value is ACCESS_DENIED, Macie doesn't have permission to retrieve the information. For example, the bucket has a restrictive bucket policy and Amazon S3 denied the request. If this value is null, Macie was able to retrieve and process the information.

bucketMetadata_jobDetails :: Lens' BucketMetadata (Maybe JobDetails)

Specifies whether any one-time or recurring classification jobs are configured to analyze data in the bucket, and, if so, the details of the job that ran most recently.

bucketMetadata_region :: Lens' BucketMetadata (Maybe Text)

The Amazon Web Services Region that hosts the bucket.

bucketMetadata_bucketCreatedAt :: Lens' BucketMetadata (Maybe UTCTime)

The date and time, in UTC and extended ISO 8601 format, when the bucket was created.

bucketMetadata_errorMessage :: Lens' BucketMetadata (Maybe Text)

A brief description of the error (errorCode) that prevented Amazon Macie from retrieving and processing information about the bucket and the bucket's objects. This value is null if Macie was able to retrieve and process the information.

bucketMetadata_bucketArn :: Lens' BucketMetadata (Maybe Text)

The Amazon Resource Name (ARN) of the bucket.

bucketMetadata_objectCount :: Lens' BucketMetadata (Maybe Integer)

The total number of objects in the bucket.

bucketMetadata_replicationDetails :: Lens' BucketMetadata (Maybe ReplicationDetails)

Specifies whether the bucket is configured to replicate one or more objects to buckets for other Amazon Web Services accounts and, if so, which accounts.

bucketMetadata_serverSideEncryption :: Lens' BucketMetadata (Maybe BucketServerSideEncryption)

Specifies whether the bucket encrypts new objects by default and, if so, the type of server-side encryption that's used.

bucketMetadata_tags :: Lens' BucketMetadata (Maybe [KeyValuePair])

An array that specifies the tags (keys and values) that are associated with the bucket.

bucketMetadata_classifiableSizeInBytes :: Lens' BucketMetadata (Maybe Integer)

The total storage size, in bytes, of the objects that Amazon Macie can analyze in the bucket. These objects use a supported storage class and have a file name extension for a supported file or storage format.

If versioning is enabled for the bucket, Macie calculates this value based on the size of the latest version of each applicable object in the bucket. This value doesn't reflect the storage size of all versions of each applicable object in the bucket.