Amazonka.MacieV2.Types.JobScheduleFrequency

Description

Documentation

data JobScheduleFrequency

Specifies the recurrence pattern for running a classification job.

See: newJobScheduleFrequency smart constructor.

Constructors

JobScheduleFrequency'

Fields

dailySchedule :: Maybe DailySchedule

Specifies a daily recurrence pattern for running the job.

monthlySchedule :: Maybe MonthlySchedule

Specifies a monthly recurrence pattern for running the job.

weeklySchedule :: Maybe WeeklySchedule

Specifies a weekly recurrence pattern for running the job.

newJobScheduleFrequency :: JobScheduleFrequency

Create a value of JobScheduleFrequency with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:dailySchedule:JobScheduleFrequency', jobScheduleFrequency_dailySchedule - Specifies a daily recurrence pattern for running the job.

$sel:monthlySchedule:JobScheduleFrequency', jobScheduleFrequency_monthlySchedule - Specifies a monthly recurrence pattern for running the job.

$sel:weeklySchedule:JobScheduleFrequency', jobScheduleFrequency_weeklySchedule - Specifies a weekly recurrence pattern for running the job.

jobScheduleFrequency_dailySchedule :: Lens' JobScheduleFrequency (Maybe DailySchedule)

Specifies a daily recurrence pattern for running the job.

jobScheduleFrequency_monthlySchedule :: Lens' JobScheduleFrequency (Maybe MonthlySchedule)

Specifies a monthly recurrence pattern for running the job.

jobScheduleFrequency_weeklySchedule :: Lens' JobScheduleFrequency (Maybe WeeklySchedule)

Specifies a weekly recurrence pattern for running the job.