Amazonka.MacieV2.Types.JobScopeTerm

Description

Documentation

data JobScopeTerm

Specifies a property- or tag-based condition that defines criteria for including or excluding S3 objects from a classification job. A JobScopeTerm object can contain only one simpleScopeTerm object or one tagScopeTerm object.

See: newJobScopeTerm smart constructor.

Constructors

JobScopeTerm'

Fields

simpleScopeTerm :: Maybe SimpleScopeTerm

A property-based condition that defines a property, operator, and one or more values for including or excluding objects from the job.

tagScopeTerm :: Maybe TagScopeTerm

A tag-based condition that defines the operator and tag keys or tag key and value pairs for including or excluding objects from the job.

newJobScopeTerm :: JobScopeTerm

Create a value of JobScopeTerm with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:simpleScopeTerm:JobScopeTerm', jobScopeTerm_simpleScopeTerm - A property-based condition that defines a property, operator, and one or more values for including or excluding objects from the job.

$sel:tagScopeTerm:JobScopeTerm', jobScopeTerm_tagScopeTerm - A tag-based condition that defines the operator and tag keys or tag key and value pairs for including or excluding objects from the job.

jobScopeTerm_simpleScopeTerm :: Lens' JobScopeTerm (Maybe SimpleScopeTerm)

A property-based condition that defines a property, operator, and one or more values for including or excluding objects from the job.

jobScopeTerm_tagScopeTerm :: Lens' JobScopeTerm (Maybe TagScopeTerm)

A tag-based condition that defines the operator and tag keys or tag key and value pairs for including or excluding objects from the job.