Amazonka.MacieV2.Types.ObjectCountByEncryptionType

Description

Documentation

data ObjectCountByEncryptionType

Provides information about the number of objects that are in an S3 bucket and use certain types of server-side encryption, use client-side encryption, or aren't encrypted.

See: newObjectCountByEncryptionType smart constructor.

Constructors

ObjectCountByEncryptionType'

Fields

unknown :: Maybe Integer

The total number of objects that Amazon Macie doesn't have current encryption metadata for. Macie can't provide current data about the encryption settings for these objects.

s3Managed :: Maybe Integer

The total number of objects that are encrypted with an Amazon S3 managed key. The objects use Amazon S3 managed encryption (SSE-S3).

unencrypted :: Maybe Integer

The total number of objects that aren't encrypted or use client-side encryption.

kmsManaged :: Maybe Integer

The total number of objects that are encrypted with an KMS key, either an Amazon Web Services managed key or a customer managed key. The objects use KMS encryption (SSE-KMS).

customerManaged :: Maybe Integer

The total number of objects that are encrypted with a customer-provided key. The objects use customer-provided server-side encryption (SSE-C).

newObjectCountByEncryptionType :: ObjectCountByEncryptionType

Create a value of ObjectCountByEncryptionType with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:unknown:ObjectCountByEncryptionType', objectCountByEncryptionType_unknown - The total number of objects that Amazon Macie doesn't have current encryption metadata for. Macie can't provide current data about the encryption settings for these objects.

$sel:s3Managed:ObjectCountByEncryptionType', objectCountByEncryptionType_s3Managed - The total number of objects that are encrypted with an Amazon S3 managed key. The objects use Amazon S3 managed encryption (SSE-S3).

$sel:unencrypted:ObjectCountByEncryptionType', objectCountByEncryptionType_unencrypted - The total number of objects that aren't encrypted or use client-side encryption.

$sel:kmsManaged:ObjectCountByEncryptionType', objectCountByEncryptionType_kmsManaged - The total number of objects that are encrypted with an KMS key, either an Amazon Web Services managed key or a customer managed key. The objects use KMS encryption (SSE-KMS).

$sel:customerManaged:ObjectCountByEncryptionType', objectCountByEncryptionType_customerManaged - The total number of objects that are encrypted with a customer-provided key. The objects use customer-provided server-side encryption (SSE-C).

objectCountByEncryptionType_unknown :: Lens' ObjectCountByEncryptionType (Maybe Integer)

The total number of objects that Amazon Macie doesn't have current encryption metadata for. Macie can't provide current data about the encryption settings for these objects.

objectCountByEncryptionType_s3Managed :: Lens' ObjectCountByEncryptionType (Maybe Integer)

The total number of objects that are encrypted with an Amazon S3 managed key. The objects use Amazon S3 managed encryption (SSE-S3).

objectCountByEncryptionType_unencrypted :: Lens' ObjectCountByEncryptionType (Maybe Integer)

The total number of objects that aren't encrypted or use client-side encryption.

objectCountByEncryptionType_kmsManaged :: Lens' ObjectCountByEncryptionType (Maybe Integer)

The total number of objects that are encrypted with an KMS key, either an Amazon Web Services managed key or a customer managed key. The objects use KMS encryption (SSE-KMS).

objectCountByEncryptionType_customerManaged :: Lens' ObjectCountByEncryptionType (Maybe Integer)

The total number of objects that are encrypted with a customer-provided key. The objects use customer-provided server-side encryption (SSE-C).