Amazonka.MacieV2.Types.Severity

Description

Documentation

data Severity

Provides the numerical and qualitative representations of a finding's severity.

See: newSeverity smart constructor.

Constructors

Severity'

Fields

score :: Maybe Integer

The numerical representation of the finding's severity, ranging from 1 (least severe) to 3 (most severe).

description :: Maybe SeverityDescription

The qualitative representation of the finding's severity, ranging from Low (least severe) to High (most severe).

newSeverity :: Severity

Create a value of Severity with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:score:Severity', severity_score - The numerical representation of the finding's severity, ranging from 1 (least severe) to 3 (most severe).

$sel:description:Severity', severity_description - The qualitative representation of the finding's severity, ranging from Low (least severe) to High (most severe).

severity_score :: Lens' Severity (Maybe Integer)

The numerical representation of the finding's severity, ranging from 1 (least severe) to 3 (most severe).

severity_description :: Lens' Severity (Maybe SeverityDescription)

The qualitative representation of the finding's severity, ranging from Low (least severe) to High (most severe).