Amazonka.MacieV2.Types.SimpleScopeTerm

Description

Documentation

data SimpleScopeTerm

Specifies a property-based condition that determines whether an S3 object is included or excluded from a classification job.

See: newSimpleScopeTerm smart constructor.

Constructors

SimpleScopeTerm'

Fields

values :: Maybe [Text]

An array that lists the values to use in the condition. If the value for the key property is OBJECT_EXTENSION or OBJECT_KEY, this array can specify multiple values and Amazon Macie uses OR logic to join the values. Otherwise, this array can specify only one value.

Valid values for each supported property (key) are:

OBJECT_EXTENSION - A string that represents the file name extension of an object. For example: docx or pdf

OBJECT_KEY - A string that represents the key prefix (folder name or path) of an object. For example: logs or awslogs/eventlogs. This value applies a condition to objects whose keys (names) begin with the specified value.

OBJECT_LAST_MODIFIED_DATE - The date and time (in UTC and extended ISO 8601 format) when an object was created or last changed, whichever is latest. For example: 2020-09-28T14:31:13Z

OBJECT_SIZE - An integer that represents the storage size (in bytes) of an object.

Macie doesn't support use of wildcard characters in these values. Also, string values are case sensitive.

key :: Maybe ScopeFilterKey

The object property to use in the condition.

comparator :: Maybe JobComparator

The operator to use in the condition. Valid values for each supported property (key) are:

OBJECT_EXTENSION - EQ (equals) or NE (not equals)

OBJECT_KEY - STARTS_WITH

OBJECT_LAST_MODIFIED_DATE - Any operator except CONTAINS

OBJECT_SIZE - Any operator except CONTAINS

newSimpleScopeTerm :: SimpleScopeTerm

Create a value of SimpleScopeTerm with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:values:SimpleScopeTerm', simpleScopeTerm_values - An array that lists the values to use in the condition. If the value for the key property is OBJECT_EXTENSION or OBJECT_KEY, this array can specify multiple values and Amazon Macie uses OR logic to join the values. Otherwise, this array can specify only one value.

Valid values for each supported property (key) are:

OBJECT_EXTENSION - A string that represents the file name extension of an object. For example: docx or pdf

OBJECT_KEY - A string that represents the key prefix (folder name or path) of an object. For example: logs or awslogs/eventlogs. This value applies a condition to objects whose keys (names) begin with the specified value.

OBJECT_LAST_MODIFIED_DATE - The date and time (in UTC and extended ISO 8601 format) when an object was created or last changed, whichever is latest. For example: 2020-09-28T14:31:13Z

OBJECT_SIZE - An integer that represents the storage size (in bytes) of an object.

Macie doesn't support use of wildcard characters in these values. Also, string values are case sensitive.

$sel:key:SimpleScopeTerm', simpleScopeTerm_key - The object property to use in the condition.

$sel:comparator:SimpleScopeTerm', simpleScopeTerm_comparator - The operator to use in the condition. Valid values for each supported property (key) are:

OBJECT_EXTENSION - EQ (equals) or NE (not equals)

OBJECT_KEY - STARTS_WITH

OBJECT_LAST_MODIFIED_DATE - Any operator except CONTAINS

OBJECT_SIZE - Any operator except CONTAINS

simpleScopeTerm_values :: Lens' SimpleScopeTerm (Maybe [Text])

An array that lists the values to use in the condition. If the value for the key property is OBJECT_EXTENSION or OBJECT_KEY, this array can specify multiple values and Amazon Macie uses OR logic to join the values. Otherwise, this array can specify only one value.

Valid values for each supported property (key) are:

OBJECT_EXTENSION - A string that represents the file name extension of an object. For example: docx or pdf

OBJECT_KEY - A string that represents the key prefix (folder name or path) of an object. For example: logs or awslogs/eventlogs. This value applies a condition to objects whose keys (names) begin with the specified value.

OBJECT_LAST_MODIFIED_DATE - The date and time (in UTC and extended ISO 8601 format) when an object was created or last changed, whichever is latest. For example: 2020-09-28T14:31:13Z

OBJECT_SIZE - An integer that represents the storage size (in bytes) of an object.

Macie doesn't support use of wildcard characters in these values. Also, string values are case sensitive.

simpleScopeTerm_key :: Lens' SimpleScopeTerm (Maybe ScopeFilterKey)

The object property to use in the condition.

simpleScopeTerm_comparator :: Lens' SimpleScopeTerm (Maybe JobComparator)

The operator to use in the condition. Valid values for each supported property (key) are:

OBJECT_EXTENSION - EQ (equals) or NE (not equals)

OBJECT_KEY - STARTS_WITH

OBJECT_LAST_MODIFIED_DATE - Any operator except CONTAINS

OBJECT_SIZE - Any operator except CONTAINS