Amazonka.MacieV2.Types.UserIdentity

Description

Documentation

data UserIdentity

Provides information about the type and other characteristics of an entity that performed an action on an affected resource.

See: newUserIdentity smart constructor.

Constructors

UserIdentity'

Fields

iamUser :: Maybe IamUser

If the action was performed using the credentials for an Identity and Access Management (IAM) user, the name and other details about the user.

root :: Maybe UserIdentityRoot

If the action was performed using the credentials for your Amazon Web Services account, the details of your account.

awsAccount :: Maybe AwsAccount

If the action was performed using the credentials for another Amazon Web Services account, the details of that account.

assumedRole :: Maybe AssumedRole

If the action was performed with temporary security credentials that were obtained using the AssumeRole operation of the Security Token Service (STS) API, the identifiers, session context, and other details about the identity.

federatedUser :: Maybe FederatedUser

If the action was performed with temporary security credentials that were obtained using the GetFederationToken operation of the Security Token Service (STS) API, the identifiers, session context, and other details about the identity.

awsService :: Maybe AwsService

If the action was performed by an Amazon Web Services account that belongs to an Amazon Web Service, the name of the service.

type' :: Maybe UserIdentityType

The type of entity that performed the action.

newUserIdentity :: UserIdentity

Create a value of UserIdentity with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:iamUser:UserIdentity', userIdentity_iamUser - If the action was performed using the credentials for an Identity and Access Management (IAM) user, the name and other details about the user.

$sel:root:UserIdentity', userIdentity_root - If the action was performed using the credentials for your Amazon Web Services account, the details of your account.

$sel:awsAccount:UserIdentity', userIdentity_awsAccount - If the action was performed using the credentials for another Amazon Web Services account, the details of that account.

$sel:assumedRole:UserIdentity', userIdentity_assumedRole - If the action was performed with temporary security credentials that were obtained using the AssumeRole operation of the Security Token Service (STS) API, the identifiers, session context, and other details about the identity.

$sel:federatedUser:UserIdentity', userIdentity_federatedUser - If the action was performed with temporary security credentials that were obtained using the GetFederationToken operation of the Security Token Service (STS) API, the identifiers, session context, and other details about the identity.

$sel:awsService:UserIdentity', userIdentity_awsService - If the action was performed by an Amazon Web Services account that belongs to an Amazon Web Service, the name of the service.

$sel:type':UserIdentity', userIdentity_type - The type of entity that performed the action.

userIdentity_iamUser :: Lens' UserIdentity (Maybe IamUser)

If the action was performed using the credentials for an Identity and Access Management (IAM) user, the name and other details about the user.

userIdentity_root :: Lens' UserIdentity (Maybe UserIdentityRoot)

If the action was performed using the credentials for your Amazon Web Services account, the details of your account.

userIdentity_awsAccount :: Lens' UserIdentity (Maybe AwsAccount)

If the action was performed using the credentials for another Amazon Web Services account, the details of that account.

userIdentity_assumedRole :: Lens' UserIdentity (Maybe AssumedRole)

If the action was performed with temporary security credentials that were obtained using the AssumeRole operation of the Security Token Service (STS) API, the identifiers, session context, and other details about the identity.

userIdentity_federatedUser :: Lens' UserIdentity (Maybe FederatedUser)

If the action was performed with temporary security credentials that were obtained using the GetFederationToken operation of the Security Token Service (STS) API, the identifiers, session context, and other details about the identity.

userIdentity_awsService :: Lens' UserIdentity (Maybe AwsService)

If the action was performed by an Amazon Web Services account that belongs to an Amazon Web Service, the name of the service.

userIdentity_type :: Lens' UserIdentity (Maybe UserIdentityType)

The type of entity that performed the action.