Amazonka.ManagedBlockChain.CreateMember

Description

Creates a member within a Managed Blockchain network.

Applies only to Hyperledger Fabric.

Creating a Request

data CreateMember

See: newCreateMember smart constructor.

Constructors

CreateMember'

Fields

clientRequestToken :: Text

A unique, case-sensitive identifier that you provide to ensure the idempotency of the operation. An idempotent operation completes no more than one time. This identifier is required only if you make a service request directly using an HTTP client. It is generated automatically if you use an AWS SDK or the AWS CLI.

invitationId :: Text

The unique identifier of the invitation that is sent to the member to join the network.

networkId :: Text

The unique identifier of the network in which the member is created.

memberConfiguration :: MemberConfiguration

Member configuration parameters.

newCreateMember

Arguments

:: Text
$sel:clientRequestToken:CreateMember'

-> Text
$sel:invitationId:CreateMember'

-> Text
$sel:networkId:CreateMember'

-> MemberConfiguration
$sel:memberConfiguration:CreateMember'

-> CreateMember

Create a value of CreateMember with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:clientRequestToken:CreateMember', createMember_clientRequestToken - A unique, case-sensitive identifier that you provide to ensure the idempotency of the operation. An idempotent operation completes no more than one time. This identifier is required only if you make a service request directly using an HTTP client. It is generated automatically if you use an AWS SDK or the AWS CLI.

$sel:invitationId:CreateMember', createMember_invitationId - The unique identifier of the invitation that is sent to the member to join the network.

$sel:networkId:CreateMember', createMember_networkId - The unique identifier of the network in which the member is created.

$sel:memberConfiguration:CreateMember', createMember_memberConfiguration - Member configuration parameters.

Request Lenses

createMember_clientRequestToken :: Lens' CreateMember Text

A unique, case-sensitive identifier that you provide to ensure the idempotency of the operation. An idempotent operation completes no more than one time. This identifier is required only if you make a service request directly using an HTTP client. It is generated automatically if you use an AWS SDK or the AWS CLI.

createMember_invitationId :: Lens' CreateMember Text

The unique identifier of the invitation that is sent to the member to join the network.

createMember_networkId :: Lens' CreateMember Text

The unique identifier of the network in which the member is created.

createMember_memberConfiguration :: Lens' CreateMember MemberConfiguration

Member configuration parameters.

Destructuring the Response

data CreateMemberResponse

See: newCreateMemberResponse smart constructor.

Constructors

CreateMemberResponse'

Fields

memberId :: Maybe Text

The unique identifier of the member.

httpStatus :: Int

The response's http status code.

newCreateMemberResponse

Arguments

:: Int
$sel:httpStatus:CreateMemberResponse'

-> CreateMemberResponse

Create a value of CreateMemberResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:memberId:CreateMemberResponse', createMemberResponse_memberId - The unique identifier of the member.

$sel:httpStatus:CreateMemberResponse', createMemberResponse_httpStatus - The response's http status code.

Response Lenses

createMemberResponse_memberId :: Lens' CreateMemberResponse (Maybe Text)

The unique identifier of the member.

createMemberResponse_httpStatus :: Lens' CreateMemberResponse Int

The response's http status code.