Amazonka.ManagedBlockChain.CreateNetwork

Description

Creates a new blockchain network using Amazon Managed Blockchain.

Applies only to Hyperledger Fabric.

Creating a Request

data CreateNetwork

See: newCreateNetwork smart constructor.

Constructors

CreateNetwork'

Fields

frameworkConfiguration :: Maybe NetworkFrameworkConfiguration

Configuration properties of the blockchain framework relevant to the network configuration.

description :: Maybe Text

An optional description for the network.

tags :: Maybe (HashMap Text Text)

Tags to assign to the network. Each tag consists of a key and optional value.

When specifying tags during creation, you can specify multiple key-value pairs in a single request, with an overall maximum of 50 tags added to each resource.

For more information about tags, see Tagging Resources in the Amazon Managed Blockchain Ethereum Developer Guide, or Tagging Resources in the Amazon Managed Blockchain Hyperledger Fabric Developer Guide.

clientRequestToken :: Text

A unique, case-sensitive identifier that you provide to ensure the idempotency of the operation. An idempotent operation completes no more than one time. This identifier is required only if you make a service request directly using an HTTP client. It is generated automatically if you use an AWS SDK or the AWS CLI.

name :: Text

The name of the network.

framework :: Framework

The blockchain framework that the network uses.

frameworkVersion :: Text

The version of the blockchain framework that the network uses.

votingPolicy :: VotingPolicy

The voting rules used by the network to determine if a proposal is approved.

memberConfiguration :: MemberConfiguration

Configuration properties for the first member within the network.

newCreateNetwork

Arguments

:: Text
$sel:clientRequestToken:CreateNetwork'

-> Text
$sel:name:CreateNetwork'

-> Framework
$sel:framework:CreateNetwork'

-> Text
$sel:frameworkVersion:CreateNetwork'

-> VotingPolicy
$sel:votingPolicy:CreateNetwork'

-> MemberConfiguration
$sel:memberConfiguration:CreateNetwork'

-> CreateNetwork

Create a value of CreateNetwork with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:frameworkConfiguration:CreateNetwork', createNetwork_frameworkConfiguration - Configuration properties of the blockchain framework relevant to the network configuration.

$sel:description:CreateNetwork', createNetwork_description - An optional description for the network.

$sel:tags:CreateNetwork', createNetwork_tags - Tags to assign to the network. Each tag consists of a key and optional value.

When specifying tags during creation, you can specify multiple key-value pairs in a single request, with an overall maximum of 50 tags added to each resource.

For more information about tags, see Tagging Resources in the Amazon Managed Blockchain Ethereum Developer Guide, or Tagging Resources in the Amazon Managed Blockchain Hyperledger Fabric Developer Guide.

$sel:clientRequestToken:CreateNetwork', createNetwork_clientRequestToken - A unique, case-sensitive identifier that you provide to ensure the idempotency of the operation. An idempotent operation completes no more than one time. This identifier is required only if you make a service request directly using an HTTP client. It is generated automatically if you use an AWS SDK or the AWS CLI.

$sel:name:CreateNetwork', createNetwork_name - The name of the network.

$sel:framework:CreateNetwork', createNetwork_framework - The blockchain framework that the network uses.

$sel:frameworkVersion:CreateNetwork', createNetwork_frameworkVersion - The version of the blockchain framework that the network uses.

$sel:votingPolicy:CreateNetwork', createNetwork_votingPolicy - The voting rules used by the network to determine if a proposal is approved.

$sel:memberConfiguration:CreateNetwork', createNetwork_memberConfiguration - Configuration properties for the first member within the network.

Request Lenses

createNetwork_frameworkConfiguration :: Lens' CreateNetwork (Maybe NetworkFrameworkConfiguration)

Configuration properties of the blockchain framework relevant to the network configuration.

createNetwork_description :: Lens' CreateNetwork (Maybe Text)

An optional description for the network.

createNetwork_tags :: Lens' CreateNetwork (Maybe (HashMap Text Text))

Tags to assign to the network. Each tag consists of a key and optional value.

When specifying tags during creation, you can specify multiple key-value pairs in a single request, with an overall maximum of 50 tags added to each resource.

For more information about tags, see Tagging Resources in the Amazon Managed Blockchain Ethereum Developer Guide, or Tagging Resources in the Amazon Managed Blockchain Hyperledger Fabric Developer Guide.

createNetwork_clientRequestToken :: Lens' CreateNetwork Text

A unique, case-sensitive identifier that you provide to ensure the idempotency of the operation. An idempotent operation completes no more than one time. This identifier is required only if you make a service request directly using an HTTP client. It is generated automatically if you use an AWS SDK or the AWS CLI.

createNetwork_name :: Lens' CreateNetwork Text

The name of the network.

createNetwork_framework :: Lens' CreateNetwork Framework

The blockchain framework that the network uses.

createNetwork_frameworkVersion :: Lens' CreateNetwork Text

The version of the blockchain framework that the network uses.

createNetwork_votingPolicy :: Lens' CreateNetwork VotingPolicy

The voting rules used by the network to determine if a proposal is approved.

createNetwork_memberConfiguration :: Lens' CreateNetwork MemberConfiguration

Configuration properties for the first member within the network.

Destructuring the Response

data CreateNetworkResponse

See: newCreateNetworkResponse smart constructor.

Constructors

CreateNetworkResponse'

Fields

memberId :: Maybe Text

The unique identifier for the first member within the network.

networkId :: Maybe Text

The unique identifier for the network.

httpStatus :: Int

The response's http status code.

newCreateNetworkResponse

Arguments

:: Int
$sel:httpStatus:CreateNetworkResponse'

-> CreateNetworkResponse

Create a value of CreateNetworkResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:memberId:CreateNetworkResponse', createNetworkResponse_memberId - The unique identifier for the first member within the network.

$sel:networkId:CreateNetworkResponse', createNetworkResponse_networkId - The unique identifier for the network.

$sel:httpStatus:CreateNetworkResponse', createNetworkResponse_httpStatus - The response's http status code.

Response Lenses

createNetworkResponse_memberId :: Lens' CreateNetworkResponse (Maybe Text)

The unique identifier for the first member within the network.

createNetworkResponse_networkId :: Lens' CreateNetworkResponse (Maybe Text)

The unique identifier for the network.

createNetworkResponse_httpStatus :: Lens' CreateNetworkResponse Int

The response's http status code.