Amazonka.ManagedBlockChain.DeleteMember

Description

Deletes a member. Deleting a member removes the member and all associated resources from the network. DeleteMember can only be called for a specified MemberId if the principal performing the action is associated with the AWS account that owns the member. In all other cases, the DeleteMember action is carried out as the result of an approved proposal to remove a member. If MemberId is the last member in a network specified by the last AWS account, the network is deleted also.

Applies only to Hyperledger Fabric.

Creating a Request

data DeleteMember

See: newDeleteMember smart constructor.

Constructors

DeleteMember'

Fields

networkId :: Text

The unique identifier of the network from which the member is removed.

memberId :: Text

The unique identifier of the member to remove.

newDeleteMember

Arguments

:: Text
$sel:networkId:DeleteMember'

-> Text
$sel:memberId:DeleteMember'

-> DeleteMember

Create a value of DeleteMember with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:networkId:DeleteMember', deleteMember_networkId - The unique identifier of the network from which the member is removed.

$sel:memberId:DeleteMember', deleteMember_memberId - The unique identifier of the member to remove.

Request Lenses

deleteMember_networkId :: Lens' DeleteMember Text

The unique identifier of the network from which the member is removed.

deleteMember_memberId :: Lens' DeleteMember Text

The unique identifier of the member to remove.

Destructuring the Response

data DeleteMemberResponse

See: newDeleteMemberResponse smart constructor.

Constructors

DeleteMemberResponse'

Fields

httpStatus :: Int

The response's http status code.

newDeleteMemberResponse

Arguments

:: Int
$sel:httpStatus:DeleteMemberResponse'

-> DeleteMemberResponse

Create a value of DeleteMemberResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:httpStatus:DeleteMemberResponse', deleteMemberResponse_httpStatus - The response's http status code.

Response Lenses

deleteMemberResponse_httpStatus :: Lens' DeleteMemberResponse Int

The response's http status code.