Amazonka.ManagedBlockChain.ListMembers

Description

Returns a list of the members in a network and properties of their configurations.

Applies only to Hyperledger Fabric.

Creating a Request

data ListMembers

See: newListMembers smart constructor.

Constructors

ListMembers'

Fields

status :: Maybe MemberStatus

An optional status specifier. If provided, only members currently in this status are listed.

nextToken :: Maybe Text

The pagination token that indicates the next set of results to retrieve.

name :: Maybe Text

The optional name of the member to list.

isOwned :: Maybe Bool

An optional Boolean value. If provided, the request is limited either to members that the current AWS account owns (true) or that other AWS accounts own (false). If omitted, all members are listed.

maxResults :: Maybe Natural

The maximum number of members to return in the request.

networkId :: Text

The unique identifier of the network for which to list members.

newListMembers

Arguments

:: Text
$sel:networkId:ListMembers'

-> ListMembers

Create a value of ListMembers with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:status:ListMembers', listMembers_status - An optional status specifier. If provided, only members currently in this status are listed.

$sel:nextToken:ListMembers', listMembers_nextToken - The pagination token that indicates the next set of results to retrieve.

$sel:name:ListMembers', listMembers_name - The optional name of the member to list.

$sel:isOwned:ListMembers', listMembers_isOwned - An optional Boolean value. If provided, the request is limited either to members that the current AWS account owns (true) or that other AWS accounts own (false). If omitted, all members are listed.

$sel:maxResults:ListMembers', listMembers_maxResults - The maximum number of members to return in the request.

$sel:networkId:ListMembers', listMembers_networkId - The unique identifier of the network for which to list members.

Request Lenses

listMembers_status :: Lens' ListMembers (Maybe MemberStatus)

An optional status specifier. If provided, only members currently in this status are listed.

listMembers_nextToken :: Lens' ListMembers (Maybe Text)

The pagination token that indicates the next set of results to retrieve.

listMembers_name :: Lens' ListMembers (Maybe Text)

The optional name of the member to list.

listMembers_isOwned :: Lens' ListMembers (Maybe Bool)

An optional Boolean value. If provided, the request is limited either to members that the current AWS account owns (true) or that other AWS accounts own (false). If omitted, all members are listed.

listMembers_maxResults :: Lens' ListMembers (Maybe Natural)

The maximum number of members to return in the request.

listMembers_networkId :: Lens' ListMembers Text

The unique identifier of the network for which to list members.

Destructuring the Response

data ListMembersResponse

See: newListMembersResponse smart constructor.

Constructors

ListMembersResponse'

Fields

members :: Maybe [MemberSummary]

An array of MemberSummary objects. Each object contains details about a network member.

nextToken :: Maybe Text

The pagination token that indicates the next set of results to retrieve.

httpStatus :: Int

The response's http status code.

newListMembersResponse

Arguments

:: Int
$sel:httpStatus:ListMembersResponse'

-> ListMembersResponse

Create a value of ListMembersResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:members:ListMembersResponse', listMembersResponse_members - An array of MemberSummary objects. Each object contains details about a network member.

$sel:nextToken:ListMembers', listMembersResponse_nextToken - The pagination token that indicates the next set of results to retrieve.

$sel:httpStatus:ListMembersResponse', listMembersResponse_httpStatus - The response's http status code.

Response Lenses

listMembersResponse_members :: Lens' ListMembersResponse (Maybe [MemberSummary])

An array of MemberSummary objects. Each object contains details about a network member.

listMembersResponse_nextToken :: Lens' ListMembersResponse (Maybe Text)

The pagination token that indicates the next set of results to retrieve.

listMembersResponse_httpStatus :: Lens' ListMembersResponse Int

The response's http status code.