Amazonka.ManagedBlockChain.ListNetworks

Description

Returns information about the networks in which the current AWS account participates.

Applies to Hyperledger Fabric and Ethereum.

Creating a Request

data ListNetworks

See: newListNetworks smart constructor.

Constructors

ListNetworks'

Fields

status :: Maybe NetworkStatus

An optional status specifier. If provided, only networks currently in this status are listed.

Applies only to Hyperledger Fabric.

framework :: Maybe Framework

An optional framework specifier. If provided, only networks of this framework type are listed.

nextToken :: Maybe Text

The pagination token that indicates the next set of results to retrieve.

name :: Maybe Text

The name of the network.

maxResults :: Maybe Natural

The maximum number of networks to list.

newListNetworks :: ListNetworks

Create a value of ListNetworks with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:status:ListNetworks', listNetworks_status - An optional status specifier. If provided, only networks currently in this status are listed.

Applies only to Hyperledger Fabric.

$sel:framework:ListNetworks', listNetworks_framework - An optional framework specifier. If provided, only networks of this framework type are listed.

$sel:nextToken:ListNetworks', listNetworks_nextToken - The pagination token that indicates the next set of results to retrieve.

$sel:name:ListNetworks', listNetworks_name - The name of the network.

$sel:maxResults:ListNetworks', listNetworks_maxResults - The maximum number of networks to list.

Request Lenses

listNetworks_status :: Lens' ListNetworks (Maybe NetworkStatus)

An optional status specifier. If provided, only networks currently in this status are listed.

Applies only to Hyperledger Fabric.

listNetworks_framework :: Lens' ListNetworks (Maybe Framework)

An optional framework specifier. If provided, only networks of this framework type are listed.

listNetworks_nextToken :: Lens' ListNetworks (Maybe Text)

The pagination token that indicates the next set of results to retrieve.

listNetworks_name :: Lens' ListNetworks (Maybe Text)

The name of the network.

listNetworks_maxResults :: Lens' ListNetworks (Maybe Natural)

The maximum number of networks to list.

Destructuring the Response

data ListNetworksResponse

See: newListNetworksResponse smart constructor.

Constructors

ListNetworksResponse'

Fields

networks :: Maybe [NetworkSummary]

An array of NetworkSummary objects that contain configuration properties for each network.

nextToken :: Maybe Text

The pagination token that indicates the next set of results to retrieve.

httpStatus :: Int

The response's http status code.

newListNetworksResponse

Arguments

:: Int
$sel:httpStatus:ListNetworksResponse'

-> ListNetworksResponse

Create a value of ListNetworksResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:networks:ListNetworksResponse', listNetworksResponse_networks - An array of NetworkSummary objects that contain configuration properties for each network.

$sel:nextToken:ListNetworks', listNetworksResponse_nextToken - The pagination token that indicates the next set of results to retrieve.

$sel:httpStatus:ListNetworksResponse', listNetworksResponse_httpStatus - The response's http status code.

Response Lenses

listNetworksResponse_networks :: Lens' ListNetworksResponse (Maybe [NetworkSummary])

An array of NetworkSummary objects that contain configuration properties for each network.

listNetworksResponse_nextToken :: Lens' ListNetworksResponse (Maybe Text)

The pagination token that indicates the next set of results to retrieve.

listNetworksResponse_httpStatus :: Lens' ListNetworksResponse Int

The response's http status code.