Amazonka.ManagedBlockChain.ListProposalVotes

Description

Returns the list of votes for a specified proposal, including the value of each vote and the unique identifier of the member that cast the vote.

Applies only to Hyperledger Fabric.

Creating a Request

data ListProposalVotes

See: newListProposalVotes smart constructor.

Constructors

ListProposalVotes'

Fields

nextToken :: Maybe Text

The pagination token that indicates the next set of results to retrieve.

maxResults :: Maybe Natural

The maximum number of votes to return.

networkId :: Text

The unique identifier of the network.

proposalId :: Text

The unique identifier of the proposal.

newListProposalVotes

Arguments

:: Text
$sel:networkId:ListProposalVotes'

-> Text
$sel:proposalId:ListProposalVotes'

-> ListProposalVotes

Create a value of ListProposalVotes with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:nextToken:ListProposalVotes', listProposalVotes_nextToken - The pagination token that indicates the next set of results to retrieve.

$sel:maxResults:ListProposalVotes', listProposalVotes_maxResults - The maximum number of votes to return.

$sel:networkId:ListProposalVotes', listProposalVotes_networkId - The unique identifier of the network.

$sel:proposalId:ListProposalVotes', listProposalVotes_proposalId - The unique identifier of the proposal.

Request Lenses

listProposalVotes_nextToken :: Lens' ListProposalVotes (Maybe Text)

The pagination token that indicates the next set of results to retrieve.

listProposalVotes_maxResults :: Lens' ListProposalVotes (Maybe Natural)

The maximum number of votes to return.

listProposalVotes_networkId :: Lens' ListProposalVotes Text

The unique identifier of the network.

listProposalVotes_proposalId :: Lens' ListProposalVotes Text

The unique identifier of the proposal.

Destructuring the Response

data ListProposalVotesResponse

See: newListProposalVotesResponse smart constructor.

Constructors

ListProposalVotesResponse'

Fields

nextToken :: Maybe Text

The pagination token that indicates the next set of results to retrieve.

proposalVotes :: Maybe [VoteSummary]

The list of votes.

httpStatus :: Int

The response's http status code.

newListProposalVotesResponse

Arguments

:: Int
$sel:httpStatus:ListProposalVotesResponse'

-> ListProposalVotesResponse

Create a value of ListProposalVotesResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:nextToken:ListProposalVotes', listProposalVotesResponse_nextToken - The pagination token that indicates the next set of results to retrieve.

$sel:proposalVotes:ListProposalVotesResponse', listProposalVotesResponse_proposalVotes - The list of votes.

$sel:httpStatus:ListProposalVotesResponse', listProposalVotesResponse_httpStatus - The response's http status code.

Response Lenses

listProposalVotesResponse_nextToken :: Lens' ListProposalVotesResponse (Maybe Text)

The pagination token that indicates the next set of results to retrieve.

listProposalVotesResponse_proposalVotes :: Lens' ListProposalVotesResponse (Maybe [VoteSummary])

The list of votes.

listProposalVotesResponse_httpStatus :: Lens' ListProposalVotesResponse Int

The response's http status code.