Amazonka.ManagedBlockChain.Types.Member

Description

Documentation

data Member

Member configuration properties.

Applies only to Hyperledger Fabric.

See: newMember smart constructor.

Constructors

Member'

Fields

status :: Maybe MemberStatus

The status of a member.

CREATING - The AWS account is in the process of creating a member.

AVAILABLE - The member has been created and can participate in the network.

CREATE_FAILED - The AWS account attempted to create a member and creation failed.

UPDATING - The member is in the process of being updated.

DELETING - The member and all associated resources are in the process of being deleted. Either the AWS account that owns the member deleted it, or the member is being deleted as the result of an APPROVED PROPOSAL to remove the member.

DELETED - The member can no longer participate on the network and all associated resources are deleted. Either the AWS account that owns the member deleted it, or the member is being deleted as the result of an APPROVED PROPOSAL to remove the member.

INACCESSIBLE_ENCRYPTION_KEY - The member is impaired and might not function as expected because it cannot access the specified customer managed key in AWS KMS for encryption at rest. Either the KMS key was disabled or deleted, or the grants on the key were revoked.

The effect of disabling or deleting a key, or revoking a grant is not immediate. The member resource might take some time to find that the key is inaccessible. When a resource is in this state, we recommend deleting and recreating the resource.

kmsKeyArn :: Maybe Text

The Amazon Resource Name (ARN) of the customer managed key in AWS Key Management Service (AWS KMS) that the member uses for encryption at rest. If the value of this parameter is "AWS Owned KMS Key", the member uses an AWS owned KMS key for encryption. This parameter is inherited by the nodes that this member owns.

logPublishingConfiguration :: Maybe MemberLogPublishingConfiguration

Configuration properties for logging events associated with a member.

arn :: Maybe Text

The Amazon Resource Name (ARN) of the member. For more information about ARNs and their format, see Amazon Resource Names (ARNs) in the AWS General Reference.

networkId :: Maybe Text

The unique identifier of the network to which the member belongs.

name :: Maybe Text

The name of the member.

id :: Maybe Text

The unique identifier of the member.

creationDate :: Maybe POSIX

The date and time that the member was created.

frameworkAttributes :: Maybe MemberFrameworkAttributes

Attributes relevant to a member for the blockchain framework that the Managed Blockchain network uses.

description :: Maybe Text

An optional description for the member.

tags :: Maybe (HashMap Text Text)

Tags assigned to the member. Tags consist of a key and optional value. For more information about tags, see Tagging Resources in the Amazon Managed Blockchain Hyperledger Fabric Developer Guide.

newMember :: Member

Create a value of Member with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:status:Member', member_status - The status of a member.

CREATING - The AWS account is in the process of creating a member.

AVAILABLE - The member has been created and can participate in the network.

CREATE_FAILED - The AWS account attempted to create a member and creation failed.

UPDATING - The member is in the process of being updated.

DELETING - The member and all associated resources are in the process of being deleted. Either the AWS account that owns the member deleted it, or the member is being deleted as the result of an APPROVED PROPOSAL to remove the member.

DELETED - The member can no longer participate on the network and all associated resources are deleted. Either the AWS account that owns the member deleted it, or the member is being deleted as the result of an APPROVED PROPOSAL to remove the member.

INACCESSIBLE_ENCRYPTION_KEY - The member is impaired and might not function as expected because it cannot access the specified customer managed key in AWS KMS for encryption at rest. Either the KMS key was disabled or deleted, or the grants on the key were revoked.

The effect of disabling or deleting a key, or revoking a grant is not immediate. The member resource might take some time to find that the key is inaccessible. When a resource is in this state, we recommend deleting and recreating the resource.

$sel:kmsKeyArn:Member', member_kmsKeyArn - The Amazon Resource Name (ARN) of the customer managed key in AWS Key Management Service (AWS KMS) that the member uses for encryption at rest. If the value of this parameter is "AWS Owned KMS Key", the member uses an AWS owned KMS key for encryption. This parameter is inherited by the nodes that this member owns.

$sel:logPublishingConfiguration:Member', member_logPublishingConfiguration - Configuration properties for logging events associated with a member.

$sel:arn:Member', member_arn - The Amazon Resource Name (ARN) of the member. For more information about ARNs and their format, see Amazon Resource Names (ARNs) in the AWS General Reference.

$sel:networkId:Member', member_networkId - The unique identifier of the network to which the member belongs.

$sel:name:Member', member_name - The name of the member.

$sel:id:Member', member_id - The unique identifier of the member.

$sel:creationDate:Member', member_creationDate - The date and time that the member was created.

$sel:frameworkAttributes:Member', member_frameworkAttributes - Attributes relevant to a member for the blockchain framework that the Managed Blockchain network uses.

$sel:description:Member', member_description - An optional description for the member.

$sel:tags:Member', member_tags - Tags assigned to the member. Tags consist of a key and optional value. For more information about tags, see Tagging Resources in the Amazon Managed Blockchain Hyperledger Fabric Developer Guide.

member_status :: Lens' Member (Maybe MemberStatus)

The status of a member.

CREATING - The AWS account is in the process of creating a member.

AVAILABLE - The member has been created and can participate in the network.

CREATE_FAILED - The AWS account attempted to create a member and creation failed.

UPDATING - The member is in the process of being updated.

DELETING - The member and all associated resources are in the process of being deleted. Either the AWS account that owns the member deleted it, or the member is being deleted as the result of an APPROVED PROPOSAL to remove the member.

DELETED - The member can no longer participate on the network and all associated resources are deleted. Either the AWS account that owns the member deleted it, or the member is being deleted as the result of an APPROVED PROPOSAL to remove the member.

INACCESSIBLE_ENCRYPTION_KEY - The member is impaired and might not function as expected because it cannot access the specified customer managed key in AWS KMS for encryption at rest. Either the KMS key was disabled or deleted, or the grants on the key were revoked.

The effect of disabling or deleting a key, or revoking a grant is not immediate. The member resource might take some time to find that the key is inaccessible. When a resource is in this state, we recommend deleting and recreating the resource.

member_kmsKeyArn :: Lens' Member (Maybe Text)

The Amazon Resource Name (ARN) of the customer managed key in AWS Key Management Service (AWS KMS) that the member uses for encryption at rest. If the value of this parameter is "AWS Owned KMS Key", the member uses an AWS owned KMS key for encryption. This parameter is inherited by the nodes that this member owns.

member_logPublishingConfiguration :: Lens' Member (Maybe MemberLogPublishingConfiguration)

Configuration properties for logging events associated with a member.

member_arn :: Lens' Member (Maybe Text)

The Amazon Resource Name (ARN) of the member. For more information about ARNs and their format, see Amazon Resource Names (ARNs) in the AWS General Reference.

member_networkId :: Lens' Member (Maybe Text)

The unique identifier of the network to which the member belongs.

member_name :: Lens' Member (Maybe Text)

The name of the member.

member_id :: Lens' Member (Maybe Text)

The unique identifier of the member.

member_creationDate :: Lens' Member (Maybe UTCTime)

The date and time that the member was created.

member_frameworkAttributes :: Lens' Member (Maybe MemberFrameworkAttributes)

Attributes relevant to a member for the blockchain framework that the Managed Blockchain network uses.

member_description :: Lens' Member (Maybe Text)

An optional description for the member.

member_tags :: Lens' Member (Maybe (HashMap Text Text))

Tags assigned to the member. Tags consist of a key and optional value. For more information about tags, see Tagging Resources in the Amazon Managed Blockchain Hyperledger Fabric Developer Guide.