Amazonka.ManagedBlockChain.Types.NetworkFabricConfiguration

Description

Documentation

data NetworkFabricConfiguration

Hyperledger Fabric configuration properties for the network.

See: newNetworkFabricConfiguration smart constructor.

Constructors

NetworkFabricConfiguration'

Fields

edition :: Edition

The edition of Amazon Managed Blockchain that the network uses. For more information, see Amazon Managed Blockchain Pricing.

newNetworkFabricConfiguration

Arguments

:: Edition
$sel:edition:NetworkFabricConfiguration'

-> NetworkFabricConfiguration

Create a value of NetworkFabricConfiguration with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:edition:NetworkFabricConfiguration', networkFabricConfiguration_edition - The edition of Amazon Managed Blockchain that the network uses. For more information, see Amazon Managed Blockchain Pricing.

networkFabricConfiguration_edition :: Lens' NetworkFabricConfiguration Edition

The edition of Amazon Managed Blockchain that the network uses. For more information, see Amazon Managed Blockchain Pricing.