Amazonka.ManagedBlockChain.Types.NetworkFrameworkConfiguration

Description

Documentation

data NetworkFrameworkConfiguration

Configuration properties relevant to the network for the blockchain framework that the network uses.

See: newNetworkFrameworkConfiguration smart constructor.

Constructors

NetworkFrameworkConfiguration'

Fields

fabric :: Maybe NetworkFabricConfiguration

Hyperledger Fabric configuration properties for a Managed Blockchain network that uses Hyperledger Fabric.

newNetworkFrameworkConfiguration :: NetworkFrameworkConfiguration

Create a value of NetworkFrameworkConfiguration with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:fabric:NetworkFrameworkConfiguration', networkFrameworkConfiguration_fabric - Hyperledger Fabric configuration properties for a Managed Blockchain network that uses Hyperledger Fabric.

networkFrameworkConfiguration_fabric :: Lens' NetworkFrameworkConfiguration (Maybe NetworkFabricConfiguration)

Hyperledger Fabric configuration properties for a Managed Blockchain network that uses Hyperledger Fabric.