Amazonka.ManagedBlockChain.Types.NetworkSummary

Description

Documentation

data NetworkSummary

A summary of network configuration properties.

See: newNetworkSummary smart constructor.

Constructors

NetworkSummary'

Fields

status :: Maybe NetworkStatus

The current status of the network.

framework :: Maybe Framework

The blockchain framework that the network uses.

arn :: Maybe Text

The Amazon Resource Name (ARN) of the network. For more information about ARNs and their format, see Amazon Resource Names (ARNs) in the AWS General Reference.

frameworkVersion :: Maybe Text

The version of the blockchain framework that the network uses.

name :: Maybe Text

The name of the network.

id :: Maybe Text

The unique identifier of the network.

creationDate :: Maybe POSIX

The date and time that the network was created.

description :: Maybe Text

An optional description of the network.

newNetworkSummary :: NetworkSummary

Create a value of NetworkSummary with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:status:NetworkSummary', networkSummary_status - The current status of the network.

$sel:framework:NetworkSummary', networkSummary_framework - The blockchain framework that the network uses.

$sel:arn:NetworkSummary', networkSummary_arn - The Amazon Resource Name (ARN) of the network. For more information about ARNs and their format, see Amazon Resource Names (ARNs) in the AWS General Reference.

$sel:frameworkVersion:NetworkSummary', networkSummary_frameworkVersion - The version of the blockchain framework that the network uses.

$sel:name:NetworkSummary', networkSummary_name - The name of the network.

$sel:id:NetworkSummary', networkSummary_id - The unique identifier of the network.

$sel:creationDate:NetworkSummary', networkSummary_creationDate - The date and time that the network was created.

$sel:description:NetworkSummary', networkSummary_description - An optional description of the network.

networkSummary_status :: Lens' NetworkSummary (Maybe NetworkStatus)

The current status of the network.

networkSummary_framework :: Lens' NetworkSummary (Maybe Framework)

The blockchain framework that the network uses.

networkSummary_arn :: Lens' NetworkSummary (Maybe Text)

The Amazon Resource Name (ARN) of the network. For more information about ARNs and their format, see Amazon Resource Names (ARNs) in the AWS General Reference.

networkSummary_frameworkVersion :: Lens' NetworkSummary (Maybe Text)

The version of the blockchain framework that the network uses.

networkSummary_name :: Lens' NetworkSummary (Maybe Text)

The name of the network.

networkSummary_id :: Lens' NetworkSummary (Maybe Text)

The unique identifier of the network.

networkSummary_creationDate :: Lens' NetworkSummary (Maybe UTCTime)

The date and time that the network was created.

networkSummary_description :: Lens' NetworkSummary (Maybe Text)

An optional description of the network.