Amazonka.ManagedBlockChain.Types.ProposalSummary

Description

Documentation

data ProposalSummary

Properties of a proposal.

Applies only to Hyperledger Fabric.

See: newProposalSummary smart constructor.

Constructors

ProposalSummary'

Fields

status :: Maybe ProposalStatus

The status of the proposal. Values are as follows:

IN_PROGRESS - The proposal is active and open for member voting.

APPROVED - The proposal was approved with sufficient YES votes among members according to the VotingPolicy specified for the Network. The specified proposal actions are carried out.

REJECTED - The proposal was rejected with insufficient YES votes among members according to the VotingPolicy specified for the Network. The specified ProposalActions are not carried out.

EXPIRED - Members did not cast the number of votes required to determine the proposal outcome before the proposal expired. The specified ProposalActions are not carried out.

ACTION_FAILED - One or more of the specified ProposalActions in a proposal that was approved could not be completed because of an error.

arn :: Maybe Text

The Amazon Resource Name (ARN) of the proposal. For more information about ARNs and their format, see Amazon Resource Names (ARNs) in the AWS General Reference.

proposedByMemberId :: Maybe Text

The unique identifier of the member that created the proposal.

proposalId :: Maybe Text

The unique identifier of the proposal.

proposedByMemberName :: Maybe Text

The name of the member that created the proposal.

expirationDate :: Maybe POSIX

The date and time that the proposal expires. This is the CreationDate plus the ProposalDurationInHours that is specified in the ProposalThresholdPolicy. After this date and time, if members have not cast enough votes to determine the outcome according to the voting policy, the proposal is EXPIRED and Actions are not carried out.

creationDate :: Maybe POSIX

The date and time that the proposal was created.

description :: Maybe Text

The description of the proposal.

newProposalSummary :: ProposalSummary

Create a value of ProposalSummary with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:status:ProposalSummary', proposalSummary_status - The status of the proposal. Values are as follows:

IN_PROGRESS - The proposal is active and open for member voting.

APPROVED - The proposal was approved with sufficient YES votes among members according to the VotingPolicy specified for the Network. The specified proposal actions are carried out.

REJECTED - The proposal was rejected with insufficient YES votes among members according to the VotingPolicy specified for the Network. The specified ProposalActions are not carried out.

EXPIRED - Members did not cast the number of votes required to determine the proposal outcome before the proposal expired. The specified ProposalActions are not carried out.

ACTION_FAILED - One or more of the specified ProposalActions in a proposal that was approved could not be completed because of an error.

$sel:arn:ProposalSummary', proposalSummary_arn - The Amazon Resource Name (ARN) of the proposal. For more information about ARNs and their format, see Amazon Resource Names (ARNs) in the AWS General Reference.

$sel:proposedByMemberId:ProposalSummary', proposalSummary_proposedByMemberId - The unique identifier of the member that created the proposal.

$sel:proposalId:ProposalSummary', proposalSummary_proposalId - The unique identifier of the proposal.

$sel:proposedByMemberName:ProposalSummary', proposalSummary_proposedByMemberName - The name of the member that created the proposal.

$sel:expirationDate:ProposalSummary', proposalSummary_expirationDate - The date and time that the proposal expires. This is the CreationDate plus the ProposalDurationInHours that is specified in the ProposalThresholdPolicy. After this date and time, if members have not cast enough votes to determine the outcome according to the voting policy, the proposal is EXPIRED and Actions are not carried out.

$sel:creationDate:ProposalSummary', proposalSummary_creationDate - The date and time that the proposal was created.

$sel:description:ProposalSummary', proposalSummary_description - The description of the proposal.

proposalSummary_status :: Lens' ProposalSummary (Maybe ProposalStatus)

The status of the proposal. Values are as follows:

IN_PROGRESS - The proposal is active and open for member voting.

APPROVED - The proposal was approved with sufficient YES votes among members according to the VotingPolicy specified for the Network. The specified proposal actions are carried out.

REJECTED - The proposal was rejected with insufficient YES votes among members according to the VotingPolicy specified for the Network. The specified ProposalActions are not carried out.

EXPIRED - Members did not cast the number of votes required to determine the proposal outcome before the proposal expired. The specified ProposalActions are not carried out.

ACTION_FAILED - One or more of the specified ProposalActions in a proposal that was approved could not be completed because of an error.

proposalSummary_arn :: Lens' ProposalSummary (Maybe Text)

The Amazon Resource Name (ARN) of the proposal. For more information about ARNs and their format, see Amazon Resource Names (ARNs) in the AWS General Reference.

proposalSummary_proposedByMemberId :: Lens' ProposalSummary (Maybe Text)

The unique identifier of the member that created the proposal.

proposalSummary_proposalId :: Lens' ProposalSummary (Maybe Text)

The unique identifier of the proposal.

proposalSummary_proposedByMemberName :: Lens' ProposalSummary (Maybe Text)

The name of the member that created the proposal.

proposalSummary_expirationDate :: Lens' ProposalSummary (Maybe UTCTime)

The date and time that the proposal expires. This is the CreationDate plus the ProposalDurationInHours that is specified in the ProposalThresholdPolicy. After this date and time, if members have not cast enough votes to determine the outcome according to the voting policy, the proposal is EXPIRED and Actions are not carried out.

proposalSummary_creationDate :: Lens' ProposalSummary (Maybe UTCTime)

The date and time that the proposal was created.

proposalSummary_description :: Lens' ProposalSummary (Maybe Text)

The description of the proposal.