Amazonka.MarketplaceCatalog

Description

Derived from API version 2018-09-17 of the AWS service descriptions, licensed under Apache 2.0.

Catalog API actions allow you to manage your entities through list, describe, and update capabilities. An entity can be a product or an offer on AWS Marketplace.

You can automate your entity update process by integrating the AWS Marketplace Catalog API with your AWS Marketplace product build or deployment pipelines. You can also create your own applications on top of the Catalog API to manage your products on AWS Marketplace.

Service Configuration

defaultService :: Service

API version 2018-09-17 of the Amazon Marketplace Catalog Service SDK configuration.

Errors

Error matchers are designed for use with the functions provided by Control.Exception.Lens. This allows catching (and rethrowing) service specific errors returned by MarketplaceCatalog.

ValidationException

_ValidationException :: AsError a => Getting (First ServiceError) a ServiceError

An error occurred during validation.

AccessDeniedException

_AccessDeniedException :: AsError a => Getting (First ServiceError) a ServiceError

Access is denied.

ServiceQuotaExceededException

_ServiceQuotaExceededException :: AsError a => Getting (First ServiceError) a ServiceError

The maximum number of open requests per account has been exceeded.

ThrottlingException

_ThrottlingException :: AsError a => Getting (First ServiceError) a ServiceError

Too many requests.

InternalServiceException

_InternalServiceException :: AsError a => Getting (First ServiceError) a ServiceError

There was an internal service exception.

ResourceNotSupportedException

_ResourceNotSupportedException :: AsError a => Getting (First ServiceError) a ServiceError

Currently, the specified resource is not supported.

ResourceNotFoundException

_ResourceNotFoundException :: AsError a => Getting (First ServiceError) a ServiceError

The specified resource wasn't found.

ResourceInUseException

_ResourceInUseException :: AsError a => Getting (First ServiceError) a ServiceError

The resource is currently in use.

Waiters

Waiters poll by repeatedly sending a request until some remote success condition configured by the Wait specification is fulfilled. The Wait specification determines how many attempts should be made, in addition to delay and retry strategies.

Operations

Some AWS operations return results that are incomplete and require subsequent requests in order to obtain the entire result set. The process of sending subsequent requests to continue where a previous request left off is called pagination. For example, the ListObjects operation of Amazon S3 returns up to 1000 objects at a time, and you must send subsequent requests with the appropriate Marker in order to retrieve the next page of results.

Operations that have an AWSPager instance can transparently perform subsequent requests, correctly setting Markers and other request facets to iterate through the entire result set of a truncated API operation. Operations which support this have an additional note in the documentation.

Many operations have the ability to filter results on the server side. See the individual operation parameters for details.

ListEntities

data ListEntities

See: newListEntities smart constructor.

Constructors

ListEntities' (Maybe Text) (Maybe (NonEmpty Filter)) (Maybe Sort) (Maybe Natural) Text Text

newListEntities

Arguments

:: Text
$sel:catalog:ListEntities'

-> Text
$sel:entityType:ListEntities'

-> ListEntities

Create a value of ListEntities with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:nextToken:ListEntities', listEntities_nextToken - The value of the next token, if it exists. Null if there are no more results.

$sel:filterList:ListEntities', listEntities_filterList - An array of filter objects. Each filter object contains two attributes, filterName and filterValues.

$sel:sort:ListEntities', listEntities_sort - An object that contains two attributes, SortBy and SortOrder.

$sel:maxResults:ListEntities', listEntities_maxResults - Specifies the upper limit of the elements on a single page. If a value isn't provided, the default value is 20.

$sel:catalog:ListEntities', listEntities_catalog - The catalog related to the request. Fixed value: AWSMarketplace

$sel:entityType:ListEntities', listEntities_entityType - The type of entities to retrieve.

data ListEntitiesResponse

See: newListEntitiesResponse smart constructor.

Constructors

ListEntitiesResponse' (Maybe [EntitySummary]) (Maybe Text) Int

newListEntitiesResponse

Arguments

:: Int
$sel:httpStatus:ListEntitiesResponse'

-> ListEntitiesResponse

Create a value of ListEntitiesResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:entitySummaryList:ListEntitiesResponse', listEntitiesResponse_entitySummaryList - Array of EntitySummary object.

$sel:nextToken:ListEntities', listEntitiesResponse_nextToken - The value of the next token if it exists. Null if there is no more result.

$sel:httpStatus:ListEntitiesResponse', listEntitiesResponse_httpStatus - The response's http status code.

ListChangeSets

data ListChangeSets

See: newListChangeSets smart constructor.

Constructors

ListChangeSets' (Maybe Text) (Maybe (NonEmpty Filter)) (Maybe Sort) (Maybe Natural) Text

newListChangeSets

Arguments

:: Text
$sel:catalog:ListChangeSets'

-> ListChangeSets

Create a value of ListChangeSets with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:nextToken:ListChangeSets', listChangeSets_nextToken - The token value retrieved from a previous call to access the next page of results.

$sel:filterList:ListChangeSets', listChangeSets_filterList - An array of filter objects.

$sel:sort:ListChangeSets', listChangeSets_sort - An object that contains two attributes, SortBy and SortOrder.

$sel:maxResults:ListChangeSets', listChangeSets_maxResults - The maximum number of results returned by a single call. This value must be provided in the next call to retrieve the next set of results. By default, this value is 20.

$sel:catalog:ListChangeSets', listChangeSets_catalog - The catalog related to the request. Fixed value: AWSMarketplace

data ListChangeSetsResponse

See: newListChangeSetsResponse smart constructor.

Constructors

ListChangeSetsResponse' (Maybe Text) (Maybe [ChangeSetSummaryListItem]) Int

newListChangeSetsResponse

Arguments

:: Int
$sel:httpStatus:ListChangeSetsResponse'

-> ListChangeSetsResponse

Create a value of ListChangeSetsResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:nextToken:ListChangeSets', listChangeSetsResponse_nextToken - The value of the next token, if it exists. Null if there are no more results.

$sel:changeSetSummaryList:ListChangeSetsResponse', listChangeSetsResponse_changeSetSummaryList - Array of ChangeSetSummaryListItem objects.

$sel:httpStatus:ListChangeSetsResponse', listChangeSetsResponse_httpStatus - The response's http status code.

StartChangeSet

data StartChangeSet

See: newStartChangeSet smart constructor.

Constructors

StartChangeSet' (Maybe Text) (Maybe Text) Text (NonEmpty Change)

newStartChangeSet

Arguments

:: Text
$sel:catalog:StartChangeSet'

-> NonEmpty Change
$sel:changeSet:StartChangeSet'

-> StartChangeSet

Create a value of StartChangeSet with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:changeSetName:StartChangeSet', startChangeSet_changeSetName - Optional case sensitive string of up to 100 ASCII characters. The change set name can be used to filter the list of change sets.

$sel:clientRequestToken:StartChangeSet', startChangeSet_clientRequestToken - A unique token to identify the request to ensure idempotency.

$sel:catalog:StartChangeSet', startChangeSet_catalog - The catalog related to the request. Fixed value: AWSMarketplace

$sel:changeSet:StartChangeSet', startChangeSet_changeSet - Array of change object.

data StartChangeSetResponse

See: newStartChangeSetResponse smart constructor.

Constructors

StartChangeSetResponse' (Maybe Text) (Maybe Text) Int

newStartChangeSetResponse

Arguments

:: Int
$sel:httpStatus:StartChangeSetResponse'

-> StartChangeSetResponse

Create a value of StartChangeSetResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:changeSetId:StartChangeSetResponse', startChangeSetResponse_changeSetId - Unique identifier generated for the request.

$sel:changeSetArn:StartChangeSetResponse', startChangeSetResponse_changeSetArn - The ARN associated to the unique identifier generated for the request.

$sel:httpStatus:StartChangeSetResponse', startChangeSetResponse_httpStatus - The response's http status code.

CancelChangeSet

data CancelChangeSet

See: newCancelChangeSet smart constructor.

Constructors

CancelChangeSet' Text Text

newCancelChangeSet

Arguments

:: Text
$sel:catalog:CancelChangeSet'

-> Text
$sel:changeSetId:CancelChangeSet'

-> CancelChangeSet

Create a value of CancelChangeSet with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:catalog:CancelChangeSet', cancelChangeSet_catalog - Required. The catalog related to the request. Fixed value: AWSMarketplace.

$sel:changeSetId:CancelChangeSet', cancelChangeSet_changeSetId - Required. The unique identifier of the StartChangeSet request that you want to cancel.

data CancelChangeSetResponse

See: newCancelChangeSetResponse smart constructor.

Constructors

CancelChangeSetResponse' (Maybe Text) (Maybe Text) Int

newCancelChangeSetResponse

Arguments

:: Int
$sel:httpStatus:CancelChangeSetResponse'

-> CancelChangeSetResponse

Create a value of CancelChangeSetResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:changeSetId:CancelChangeSet', cancelChangeSetResponse_changeSetId - The unique identifier for the change set referenced in this request.

$sel:changeSetArn:CancelChangeSetResponse', cancelChangeSetResponse_changeSetArn - The ARN associated with the change set referenced in this request.

$sel:httpStatus:CancelChangeSetResponse', cancelChangeSetResponse_httpStatus - The response's http status code.

DescribeEntity

data DescribeEntity

See: newDescribeEntity smart constructor.

Constructors

DescribeEntity' Text Text

newDescribeEntity

Arguments

:: Text
$sel:catalog:DescribeEntity'

-> Text
$sel:entityId:DescribeEntity'

-> DescribeEntity

Create a value of DescribeEntity with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:catalog:DescribeEntity', describeEntity_catalog - Required. The catalog related to the request. Fixed value: AWSMarketplace

$sel:entityId:DescribeEntity', describeEntity_entityId - Required. The unique ID of the entity to describe.

data DescribeEntityResponse

See: newDescribeEntityResponse smart constructor.

Constructors

DescribeEntityResponse' (Maybe Text) (Maybe Text) (Maybe Text) (Maybe Text) (Maybe Text) Int

newDescribeEntityResponse

Arguments

:: Int
$sel:httpStatus:DescribeEntityResponse'

-> DescribeEntityResponse

Create a value of DescribeEntityResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:lastModifiedDate:DescribeEntityResponse', describeEntityResponse_lastModifiedDate - The last modified date of the entity, in ISO 8601 format (2018-02-27T13:45:22Z).

$sel:entityType:DescribeEntityResponse', describeEntityResponse_entityType - The named type of the entity, in the format of EntityType@Version.

$sel:details:DescribeEntityResponse', describeEntityResponse_details - This stringified JSON object includes the details of the entity.

$sel:entityIdentifier:DescribeEntityResponse', describeEntityResponse_entityIdentifier - The identifier of the entity, in the format of EntityId@RevisionId.

$sel:entityArn:DescribeEntityResponse', describeEntityResponse_entityArn - The ARN associated to the unique identifier for the change set referenced in this request.

$sel:httpStatus:DescribeEntityResponse', describeEntityResponse_httpStatus - The response's http status code.

DescribeChangeSet

data DescribeChangeSet

See: newDescribeChangeSet smart constructor.

Constructors

DescribeChangeSet' Text Text

newDescribeChangeSet

Arguments

:: Text
$sel:catalog:DescribeChangeSet'

-> Text
$sel:changeSetId:DescribeChangeSet'

-> DescribeChangeSet

Create a value of DescribeChangeSet with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:catalog:DescribeChangeSet', describeChangeSet_catalog - Required. The catalog related to the request. Fixed value: AWSMarketplace

$sel:changeSetId:DescribeChangeSet', describeChangeSet_changeSetId - Required. The unique identifier for the StartChangeSet request that you want to describe the details for.

data DescribeChangeSetResponse

See: newDescribeChangeSetResponse smart constructor.

Constructors

DescribeChangeSetResponse' (Maybe ChangeStatus) (Maybe Text) (Maybe FailureCode) (Maybe Text) (Maybe Text) (Maybe Text) (Maybe [ChangeSummary]) (Maybe Text) (Maybe Text) Int

newDescribeChangeSetResponse

Arguments

:: Int
$sel:httpStatus:DescribeChangeSetResponse'

-> DescribeChangeSetResponse

Create a value of DescribeChangeSetResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:status:DescribeChangeSetResponse', describeChangeSetResponse_status - The status of the change request.

$sel:startTime:DescribeChangeSetResponse', describeChangeSetResponse_startTime - The date and time, in ISO 8601 format (2018-02-27T13:45:22Z), the request started.

$sel:failureCode:DescribeChangeSetResponse', describeChangeSetResponse_failureCode - Returned if the change set is in FAILED status. Can be either CLIENT_ERROR, which means that there are issues with the request (see the ErrorDetailList), or SERVER_FAULT, which means that there is a problem in the system, and you should retry your request.

$sel:changeSetName:DescribeChangeSetResponse', describeChangeSetResponse_changeSetName - The optional name provided in the StartChangeSet request. If you do not provide a name, one is set by default.

$sel:failureDescription:DescribeChangeSetResponse', describeChangeSetResponse_failureDescription - Returned if there is a failure on the change set, but that failure is not related to any of the changes in the request.

$sel:changeSetId:DescribeChangeSet', describeChangeSetResponse_changeSetId - Required. The unique identifier for the change set referenced in this request.

$sel:changeSet:DescribeChangeSetResponse', describeChangeSetResponse_changeSet - An array of ChangeSummary objects.

$sel:endTime:DescribeChangeSetResponse', describeChangeSetResponse_endTime - The date and time, in ISO 8601 format (2018-02-27T13:45:22Z), the request transitioned to a terminal state. The change cannot transition to a different state. Null if the request is not in a terminal state.

$sel:changeSetArn:DescribeChangeSetResponse', describeChangeSetResponse_changeSetArn - The ARN associated with the unique identifier for the change set referenced in this request.

$sel:httpStatus:DescribeChangeSetResponse', describeChangeSetResponse_httpStatus - The response's http status code.

Types

ChangeStatus

newtype ChangeStatus

Constructors

ChangeStatus'

Fields

fromChangeStatus :: Text

Bundled Patterns

pattern ChangeStatus_APPLYING :: ChangeStatus

pattern ChangeStatus_CANCELLED :: ChangeStatus

pattern ChangeStatus_FAILED :: ChangeStatus

pattern ChangeStatus_PREPARING :: ChangeStatus

pattern ChangeStatus_SUCCEEDED :: ChangeStatus

FailureCode

newtype FailureCode

Constructors

FailureCode'

Fields

fromFailureCode :: Text

Bundled Patterns

pattern FailureCode_CLIENT_ERROR :: FailureCode

pattern FailureCode_SERVER_FAULT :: FailureCode

SortOrder

newtype SortOrder

Constructors

SortOrder'

Fields

fromSortOrder :: Text

Bundled Patterns

pattern SortOrder_ASCENDING :: SortOrder

pattern SortOrder_DESCENDING :: SortOrder

Change

data Change

An object that contains the ChangeType, Details, and Entity.

See: newChange smart constructor.

Constructors

Change' (Maybe Text) Text Entity Text

newChange

Arguments

:: Text
$sel:changeType:Change'

-> Entity
$sel:entity:Change'

-> Text
$sel:details:Change'

-> Change

Create a value of Change with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:changeName:Change', change_changeName - Optional name for the change.

$sel:changeType:Change', change_changeType - Change types are single string values that describe your intention for the change. Each change type is unique for each EntityType provided in the change's scope.

$sel:entity:Change', change_entity - The entity to be changed.

$sel:details:Change', change_details - This object contains details specific to the change type of the requested change.

ChangeSetSummaryListItem

data ChangeSetSummaryListItem

A summary of a change set returned in a list of change sets when the ListChangeSets action is called.

See: newChangeSetSummaryListItem smart constructor.

Constructors

ChangeSetSummaryListItem' (Maybe ChangeStatus) (Maybe [Text]) (Maybe Text) (Maybe FailureCode) (Maybe Text) (Maybe Text) (Maybe Text) (Maybe Text)

newChangeSetSummaryListItem :: ChangeSetSummaryListItem

Create a value of ChangeSetSummaryListItem with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:status:ChangeSetSummaryListItem', changeSetSummaryListItem_status - The current status of the change set.

$sel:entityIdList:ChangeSetSummaryListItem', changeSetSummaryListItem_entityIdList - This object is a list of entity IDs (string) that are a part of a change set. The entity ID list is a maximum of 20 entities. It must contain at least one entity.

$sel:startTime:ChangeSetSummaryListItem', changeSetSummaryListItem_startTime - The time, in ISO 8601 format (2018-02-27T13:45:22Z), when the change set was started.

$sel:failureCode:ChangeSetSummaryListItem', changeSetSummaryListItem_failureCode - Returned if the change set is in FAILED status. Can be either CLIENT_ERROR, which means that there are issues with the request (see the ErrorDetailList of DescribeChangeSet), or SERVER_FAULT, which means that there is a problem in the system, and you should retry your request.

$sel:changeSetName:ChangeSetSummaryListItem', changeSetSummaryListItem_changeSetName - The non-unique name for the change set.

$sel:changeSetId:ChangeSetSummaryListItem', changeSetSummaryListItem_changeSetId - The unique identifier for a change set.

$sel:endTime:ChangeSetSummaryListItem', changeSetSummaryListItem_endTime - The time, in ISO 8601 format (2018-02-27T13:45:22Z), when the change set was finished.

$sel:changeSetArn:ChangeSetSummaryListItem', changeSetSummaryListItem_changeSetArn - The ARN associated with the unique identifier for the change set referenced in this request.

ChangeSummary

data ChangeSummary

This object is a container for common summary information about the change. The summary doesn't contain the whole change structure.

See: newChangeSummary smart constructor.

Constructors

ChangeSummary' (Maybe Text) (Maybe Text) (Maybe [ErrorDetail]) (Maybe Entity) (Maybe Text)

newChangeSummary :: ChangeSummary

Create a value of ChangeSummary with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:changeName:ChangeSummary', changeSummary_changeName - Optional name for the change.

$sel:details:ChangeSummary', changeSummary_details - This object contains details specific to the change type of the requested change.

$sel:errorDetailList:ChangeSummary', changeSummary_errorDetailList - An array of ErrorDetail objects associated with the change.

$sel:entity:ChangeSummary', changeSummary_entity - The entity to be changed.

$sel:changeType:ChangeSummary', changeSummary_changeType - The type of the change.

Entity

data Entity

An entity contains data that describes your product, its supported features, and how it can be used or launched by your customer.

See: newEntity smart constructor.

Constructors

Entity' (Maybe Text) Text

newEntity

Arguments

:: Text
$sel:type':Entity'

-> Entity

Create a value of Entity with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:identifier:Entity', entity_identifier - The identifier for the entity.

$sel:type':Entity', entity_type - The type of entity.

EntitySummary

data EntitySummary

This object is a container for common summary information about the entity. The summary doesn't contain the whole entity structure, but it does contain information common across all entities.

See: newEntitySummary smart constructor.

Constructors

EntitySummary' (Maybe Text) (Maybe Text) (Maybe Text) (Maybe Text) (Maybe Text) (Maybe Text)

newEntitySummary :: EntitySummary

Create a value of EntitySummary with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:lastModifiedDate:EntitySummary', entitySummary_lastModifiedDate - The last time the entity was published, using ISO 8601 format (2018-02-27T13:45:22Z).

$sel:entityType:EntitySummary', entitySummary_entityType - The type of the entity.

$sel:visibility:EntitySummary', entitySummary_visibility - The visibility status of the entity to buyers. This value can be Public (everyone can view the entity), Limited (the entity is visible to limited accounts only), or Restricted (the entity was published and then unpublished and only existing buyers can view it).

$sel:name:EntitySummary', entitySummary_name - The name for the entity. This value is not unique. It is defined by the seller.

$sel:entityId:EntitySummary', entitySummary_entityId - The unique identifier for the entity.

$sel:entityArn:EntitySummary', entitySummary_entityArn - The ARN associated with the unique identifier for the entity.

ErrorDetail

data ErrorDetail

Details about the error.

See: newErrorDetail smart constructor.

Constructors

ErrorDetail' (Maybe Text) (Maybe Text)

newErrorDetail :: ErrorDetail

Create a value of ErrorDetail with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:errorCode:ErrorDetail', errorDetail_errorCode - The error code that identifies the type of error.

$sel:errorMessage:ErrorDetail', errorDetail_errorMessage - The message for the error.

Filter

data Filter

A filter object, used to optionally filter results from calls to the ListEntities and ListChangeSets actions.

See: newFilter smart constructor.

Constructors

Filter' (Maybe (NonEmpty Text)) (Maybe Text)

newFilter :: Filter

Create a value of Filter with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:valueList:Filter', filter_valueList - ListEntities - This is a list of unique EntityIds.

ListChangeSets - The supported filter names and associated ValueLists is as follows:

ChangeSetName - The supported ValueList is a list of non-unique ChangeSetNames. These are defined when you call the StartChangeSet action.

Status - The supported ValueList is a list of statuses for all change set requests.

EntityId - The supported ValueList is a list of unique EntityIds.

BeforeStartTime - The supported ValueList is a list of all change sets that started before the filter value.

AfterStartTime - The supported ValueList is a list of all change sets that started after the filter value.

BeforeEndTime - The supported ValueList is a list of all change sets that ended before the filter value.

AfterEndTime - The supported ValueList is a list of all change sets that ended after the filter value.

$sel:name:Filter', filter_name - For ListEntities, the supported value for this is an EntityId.

For ListChangeSets, the supported values are as follows:

Sort

data Sort

An object that contains two attributes, SortBy and SortOrder.

See: newSort smart constructor.

Constructors

Sort' (Maybe SortOrder) (Maybe Text)

newSort :: Sort

Create a value of Sort with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:sortOrder:Sort', sort_sortOrder - The sorting order. Can be ASCENDING or DESCENDING. The default value is DESCENDING.

$sel:sortBy:Sort', sort_sortBy - For ListEntities, supported attributes include LastModifiedDate (default), Visibility, EntityId, and Name.

For ListChangeSets, supported attributes include StartTime and EndTime.