Amazonka.MarketplaceCatalog.DescribeChangeSet

Description

Provides information about a given change set.

Creating a Request

data DescribeChangeSet

See: newDescribeChangeSet smart constructor.

Constructors

DescribeChangeSet'

Fields

catalog :: Text

Required. The catalog related to the request. Fixed value: AWSMarketplace

changeSetId :: Text

Required. The unique identifier for the StartChangeSet request that you want to describe the details for.

newDescribeChangeSet

Arguments

:: Text
$sel:catalog:DescribeChangeSet'

-> Text
$sel:changeSetId:DescribeChangeSet'

-> DescribeChangeSet

Create a value of DescribeChangeSet with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:catalog:DescribeChangeSet', describeChangeSet_catalog - Required. The catalog related to the request. Fixed value: AWSMarketplace

$sel:changeSetId:DescribeChangeSet', describeChangeSet_changeSetId - Required. The unique identifier for the StartChangeSet request that you want to describe the details for.

Request Lenses

describeChangeSet_catalog :: Lens' DescribeChangeSet Text

Required. The catalog related to the request. Fixed value: AWSMarketplace

describeChangeSet_changeSetId :: Lens' DescribeChangeSet Text

Required. The unique identifier for the StartChangeSet request that you want to describe the details for.

Destructuring the Response

data DescribeChangeSetResponse

See: newDescribeChangeSetResponse smart constructor.

Constructors

DescribeChangeSetResponse'

Fields

status :: Maybe ChangeStatus

The status of the change request.

startTime :: Maybe Text

The date and time, in ISO 8601 format (2018-02-27T13:45:22Z), the request started.

failureCode :: Maybe FailureCode

Returned if the change set is in FAILED status. Can be either CLIENT_ERROR, which means that there are issues with the request (see the ErrorDetailList), or SERVER_FAULT, which means that there is a problem in the system, and you should retry your request.

changeSetName :: Maybe Text

The optional name provided in the StartChangeSet request. If you do not provide a name, one is set by default.

failureDescription :: Maybe Text

Returned if there is a failure on the change set, but that failure is not related to any of the changes in the request.

changeSetId :: Maybe Text

Required. The unique identifier for the change set referenced in this request.

changeSet :: Maybe [ChangeSummary]

An array of ChangeSummary objects.

endTime :: Maybe Text

The date and time, in ISO 8601 format (2018-02-27T13:45:22Z), the request transitioned to a terminal state. The change cannot transition to a different state. Null if the request is not in a terminal state.

changeSetArn :: Maybe Text

The ARN associated with the unique identifier for the change set referenced in this request.

httpStatus :: Int

The response's http status code.

newDescribeChangeSetResponse

Arguments

:: Int
$sel:httpStatus:DescribeChangeSetResponse'

-> DescribeChangeSetResponse

Create a value of DescribeChangeSetResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:status:DescribeChangeSetResponse', describeChangeSetResponse_status - The status of the change request.

$sel:startTime:DescribeChangeSetResponse', describeChangeSetResponse_startTime - The date and time, in ISO 8601 format (2018-02-27T13:45:22Z), the request started.

$sel:failureCode:DescribeChangeSetResponse', describeChangeSetResponse_failureCode - Returned if the change set is in FAILED status. Can be either CLIENT_ERROR, which means that there are issues with the request (see the ErrorDetailList), or SERVER_FAULT, which means that there is a problem in the system, and you should retry your request.

$sel:changeSetName:DescribeChangeSetResponse', describeChangeSetResponse_changeSetName - The optional name provided in the StartChangeSet request. If you do not provide a name, one is set by default.

$sel:failureDescription:DescribeChangeSetResponse', describeChangeSetResponse_failureDescription - Returned if there is a failure on the change set, but that failure is not related to any of the changes in the request.

$sel:changeSetId:DescribeChangeSet', describeChangeSetResponse_changeSetId - Required. The unique identifier for the change set referenced in this request.

$sel:changeSet:DescribeChangeSetResponse', describeChangeSetResponse_changeSet - An array of ChangeSummary objects.

$sel:endTime:DescribeChangeSetResponse', describeChangeSetResponse_endTime - The date and time, in ISO 8601 format (2018-02-27T13:45:22Z), the request transitioned to a terminal state. The change cannot transition to a different state. Null if the request is not in a terminal state.

$sel:changeSetArn:DescribeChangeSetResponse', describeChangeSetResponse_changeSetArn - The ARN associated with the unique identifier for the change set referenced in this request.

$sel:httpStatus:DescribeChangeSetResponse', describeChangeSetResponse_httpStatus - The response's http status code.

Response Lenses

describeChangeSetResponse_status :: Lens' DescribeChangeSetResponse (Maybe ChangeStatus)

The status of the change request.

describeChangeSetResponse_startTime :: Lens' DescribeChangeSetResponse (Maybe Text)

The date and time, in ISO 8601 format (2018-02-27T13:45:22Z), the request started.

describeChangeSetResponse_failureCode :: Lens' DescribeChangeSetResponse (Maybe FailureCode)

Returned if the change set is in FAILED status. Can be either CLIENT_ERROR, which means that there are issues with the request (see the ErrorDetailList), or SERVER_FAULT, which means that there is a problem in the system, and you should retry your request.

describeChangeSetResponse_changeSetName :: Lens' DescribeChangeSetResponse (Maybe Text)

The optional name provided in the StartChangeSet request. If you do not provide a name, one is set by default.

describeChangeSetResponse_failureDescription :: Lens' DescribeChangeSetResponse (Maybe Text)

Returned if there is a failure on the change set, but that failure is not related to any of the changes in the request.

describeChangeSetResponse_changeSetId :: Lens' DescribeChangeSetResponse (Maybe Text)

Required. The unique identifier for the change set referenced in this request.

describeChangeSetResponse_changeSet :: Lens' DescribeChangeSetResponse (Maybe [ChangeSummary])

An array of ChangeSummary objects.

describeChangeSetResponse_endTime :: Lens' DescribeChangeSetResponse (Maybe Text)

The date and time, in ISO 8601 format (2018-02-27T13:45:22Z), the request transitioned to a terminal state. The change cannot transition to a different state. Null if the request is not in a terminal state.

describeChangeSetResponse_changeSetArn :: Lens' DescribeChangeSetResponse (Maybe Text)

The ARN associated with the unique identifier for the change set referenced in this request.

describeChangeSetResponse_httpStatus :: Lens' DescribeChangeSetResponse Int

The response's http status code.