Amazonka.MarketplaceCatalog.Lens

Description

Operations

ListEntities

listEntities_nextToken :: Lens' ListEntities (Maybe Text)

The value of the next token, if it exists. Null if there are no more results.

listEntities_filterList :: Lens' ListEntities (Maybe (NonEmpty Filter))

An array of filter objects. Each filter object contains two attributes, filterName and filterValues.

listEntities_sort :: Lens' ListEntities (Maybe Sort)

An object that contains two attributes, SortBy and SortOrder.

listEntities_maxResults :: Lens' ListEntities (Maybe Natural)

Specifies the upper limit of the elements on a single page. If a value isn't provided, the default value is 20.

listEntities_catalog :: Lens' ListEntities Text

The catalog related to the request. Fixed value: AWSMarketplace

listEntities_entityType :: Lens' ListEntities Text

The type of entities to retrieve.

listEntitiesResponse_entitySummaryList :: Lens' ListEntitiesResponse (Maybe [EntitySummary])

Array of EntitySummary object.

listEntitiesResponse_nextToken :: Lens' ListEntitiesResponse (Maybe Text)

The value of the next token if it exists. Null if there is no more result.

listEntitiesResponse_httpStatus :: Lens' ListEntitiesResponse Int

The response's http status code.

ListChangeSets

listChangeSets_nextToken :: Lens' ListChangeSets (Maybe Text)

The token value retrieved from a previous call to access the next page of results.

listChangeSets_filterList :: Lens' ListChangeSets (Maybe (NonEmpty Filter))

An array of filter objects.

listChangeSets_sort :: Lens' ListChangeSets (Maybe Sort)

An object that contains two attributes, SortBy and SortOrder.

listChangeSets_maxResults :: Lens' ListChangeSets (Maybe Natural)

The maximum number of results returned by a single call. This value must be provided in the next call to retrieve the next set of results. By default, this value is 20.

listChangeSets_catalog :: Lens' ListChangeSets Text

The catalog related to the request. Fixed value: AWSMarketplace

listChangeSetsResponse_nextToken :: Lens' ListChangeSetsResponse (Maybe Text)

The value of the next token, if it exists. Null if there are no more results.

listChangeSetsResponse_changeSetSummaryList :: Lens' ListChangeSetsResponse (Maybe [ChangeSetSummaryListItem])

Array of ChangeSetSummaryListItem objects.

listChangeSetsResponse_httpStatus :: Lens' ListChangeSetsResponse Int

The response's http status code.

StartChangeSet

startChangeSet_changeSetName :: Lens' StartChangeSet (Maybe Text)

Optional case sensitive string of up to 100 ASCII characters. The change set name can be used to filter the list of change sets.

startChangeSet_clientRequestToken :: Lens' StartChangeSet (Maybe Text)

A unique token to identify the request to ensure idempotency.

startChangeSet_catalog :: Lens' StartChangeSet Text

The catalog related to the request. Fixed value: AWSMarketplace

startChangeSet_changeSet :: Lens' StartChangeSet (NonEmpty Change)

Array of change object.

startChangeSetResponse_changeSetId :: Lens' StartChangeSetResponse (Maybe Text)

Unique identifier generated for the request.

startChangeSetResponse_changeSetArn :: Lens' StartChangeSetResponse (Maybe Text)

The ARN associated to the unique identifier generated for the request.

startChangeSetResponse_httpStatus :: Lens' StartChangeSetResponse Int

The response's http status code.

CancelChangeSet

cancelChangeSet_catalog :: Lens' CancelChangeSet Text

Required. The catalog related to the request. Fixed value: AWSMarketplace.

cancelChangeSet_changeSetId :: Lens' CancelChangeSet Text

Required. The unique identifier of the StartChangeSet request that you want to cancel.

cancelChangeSetResponse_changeSetId :: Lens' CancelChangeSetResponse (Maybe Text)

The unique identifier for the change set referenced in this request.

cancelChangeSetResponse_changeSetArn :: Lens' CancelChangeSetResponse (Maybe Text)

The ARN associated with the change set referenced in this request.

cancelChangeSetResponse_httpStatus :: Lens' CancelChangeSetResponse Int

The response's http status code.

DescribeEntity

describeEntity_catalog :: Lens' DescribeEntity Text

Required. The catalog related to the request. Fixed value: AWSMarketplace

describeEntity_entityId :: Lens' DescribeEntity Text

Required. The unique ID of the entity to describe.

describeEntityResponse_lastModifiedDate :: Lens' DescribeEntityResponse (Maybe Text)

The last modified date of the entity, in ISO 8601 format (2018-02-27T13:45:22Z).

describeEntityResponse_entityType :: Lens' DescribeEntityResponse (Maybe Text)

The named type of the entity, in the format of EntityType@Version.

describeEntityResponse_details :: Lens' DescribeEntityResponse (Maybe Text)

This stringified JSON object includes the details of the entity.

describeEntityResponse_entityIdentifier :: Lens' DescribeEntityResponse (Maybe Text)

The identifier of the entity, in the format of EntityId@RevisionId.

describeEntityResponse_entityArn :: Lens' DescribeEntityResponse (Maybe Text)

The ARN associated to the unique identifier for the change set referenced in this request.

describeEntityResponse_httpStatus :: Lens' DescribeEntityResponse Int

The response's http status code.

DescribeChangeSet

describeChangeSet_catalog :: Lens' DescribeChangeSet Text

Required. The catalog related to the request. Fixed value: AWSMarketplace

describeChangeSet_changeSetId :: Lens' DescribeChangeSet Text

Required. The unique identifier for the StartChangeSet request that you want to describe the details for.

describeChangeSetResponse_status :: Lens' DescribeChangeSetResponse (Maybe ChangeStatus)

The status of the change request.

describeChangeSetResponse_startTime :: Lens' DescribeChangeSetResponse (Maybe Text)

The date and time, in ISO 8601 format (2018-02-27T13:45:22Z), the request started.

describeChangeSetResponse_failureCode :: Lens' DescribeChangeSetResponse (Maybe FailureCode)

Returned if the change set is in FAILED status. Can be either CLIENT_ERROR, which means that there are issues with the request (see the ErrorDetailList), or SERVER_FAULT, which means that there is a problem in the system, and you should retry your request.

describeChangeSetResponse_changeSetName :: Lens' DescribeChangeSetResponse (Maybe Text)

The optional name provided in the StartChangeSet request. If you do not provide a name, one is set by default.

describeChangeSetResponse_failureDescription :: Lens' DescribeChangeSetResponse (Maybe Text)

Returned if there is a failure on the change set, but that failure is not related to any of the changes in the request.

describeChangeSetResponse_changeSetId :: Lens' DescribeChangeSetResponse (Maybe Text)

Required. The unique identifier for the change set referenced in this request.

describeChangeSetResponse_changeSet :: Lens' DescribeChangeSetResponse (Maybe [ChangeSummary])

An array of ChangeSummary objects.

describeChangeSetResponse_endTime :: Lens' DescribeChangeSetResponse (Maybe Text)

The date and time, in ISO 8601 format (2018-02-27T13:45:22Z), the request transitioned to a terminal state. The change cannot transition to a different state. Null if the request is not in a terminal state.

describeChangeSetResponse_changeSetArn :: Lens' DescribeChangeSetResponse (Maybe Text)

The ARN associated with the unique identifier for the change set referenced in this request.

describeChangeSetResponse_httpStatus :: Lens' DescribeChangeSetResponse Int

The response's http status code.

Types

Change

change_changeName :: Lens' Change (Maybe Text)

Optional name for the change.

change_changeType :: Lens' Change Text

Change types are single string values that describe your intention for the change. Each change type is unique for each EntityType provided in the change's scope.

change_entity :: Lens' Change Entity

The entity to be changed.

change_details :: Lens' Change Text

This object contains details specific to the change type of the requested change.

ChangeSetSummaryListItem

changeSetSummaryListItem_status :: Lens' ChangeSetSummaryListItem (Maybe ChangeStatus)

The current status of the change set.

changeSetSummaryListItem_entityIdList :: Lens' ChangeSetSummaryListItem (Maybe [Text])

This object is a list of entity IDs (string) that are a part of a change set. The entity ID list is a maximum of 20 entities. It must contain at least one entity.

changeSetSummaryListItem_startTime :: Lens' ChangeSetSummaryListItem (Maybe Text)

The time, in ISO 8601 format (2018-02-27T13:45:22Z), when the change set was started.

changeSetSummaryListItem_failureCode :: Lens' ChangeSetSummaryListItem (Maybe FailureCode)

Returned if the change set is in FAILED status. Can be either CLIENT_ERROR, which means that there are issues with the request (see the ErrorDetailList of DescribeChangeSet), or SERVER_FAULT, which means that there is a problem in the system, and you should retry your request.

changeSetSummaryListItem_changeSetName :: Lens' ChangeSetSummaryListItem (Maybe Text)

The non-unique name for the change set.

changeSetSummaryListItem_changeSetId :: Lens' ChangeSetSummaryListItem (Maybe Text)

The unique identifier for a change set.

changeSetSummaryListItem_endTime :: Lens' ChangeSetSummaryListItem (Maybe Text)

The time, in ISO 8601 format (2018-02-27T13:45:22Z), when the change set was finished.

changeSetSummaryListItem_changeSetArn :: Lens' ChangeSetSummaryListItem (Maybe Text)

The ARN associated with the unique identifier for the change set referenced in this request.

ChangeSummary

changeSummary_changeName :: Lens' ChangeSummary (Maybe Text)

Optional name for the change.

changeSummary_details :: Lens' ChangeSummary (Maybe Text)

This object contains details specific to the change type of the requested change.

changeSummary_errorDetailList :: Lens' ChangeSummary (Maybe [ErrorDetail])

An array of ErrorDetail objects associated with the change.

changeSummary_entity :: Lens' ChangeSummary (Maybe Entity)

The entity to be changed.

changeSummary_changeType :: Lens' ChangeSummary (Maybe Text)

The type of the change.

Entity

entity_identifier :: Lens' Entity (Maybe Text)

The identifier for the entity.

entity_type :: Lens' Entity Text

The type of entity.

EntitySummary

entitySummary_lastModifiedDate :: Lens' EntitySummary (Maybe Text)

The last time the entity was published, using ISO 8601 format (2018-02-27T13:45:22Z).

entitySummary_entityType :: Lens' EntitySummary (Maybe Text)

The type of the entity.

entitySummary_visibility :: Lens' EntitySummary (Maybe Text)

The visibility status of the entity to buyers. This value can be Public (everyone can view the entity), Limited (the entity is visible to limited accounts only), or Restricted (the entity was published and then unpublished and only existing buyers can view it).

entitySummary_name :: Lens' EntitySummary (Maybe Text)

The name for the entity. This value is not unique. It is defined by the seller.

entitySummary_entityId :: Lens' EntitySummary (Maybe Text)

The unique identifier for the entity.

entitySummary_entityArn :: Lens' EntitySummary (Maybe Text)

The ARN associated with the unique identifier for the entity.

ErrorDetail

errorDetail_errorCode :: Lens' ErrorDetail (Maybe Text)

The error code that identifies the type of error.

errorDetail_errorMessage :: Lens' ErrorDetail (Maybe Text)

The message for the error.

Filter

filter_valueList :: Lens' Filter (Maybe (NonEmpty Text))

ListEntities - This is a list of unique EntityIds.

ListChangeSets - The supported filter names and associated ValueLists is as follows:

ChangeSetName - The supported ValueList is a list of non-unique ChangeSetNames. These are defined when you call the StartChangeSet action.

Status - The supported ValueList is a list of statuses for all change set requests.

EntityId - The supported ValueList is a list of unique EntityIds.

BeforeStartTime - The supported ValueList is a list of all change sets that started before the filter value.

AfterStartTime - The supported ValueList is a list of all change sets that started after the filter value.

BeforeEndTime - The supported ValueList is a list of all change sets that ended before the filter value.

AfterEndTime - The supported ValueList is a list of all change sets that ended after the filter value.

filter_name :: Lens' Filter (Maybe Text)

For ListEntities, the supported value for this is an EntityId.

For ListChangeSets, the supported values are as follows:

Sort

sort_sortOrder :: Lens' Sort (Maybe SortOrder)

The sorting order. Can be ASCENDING or DESCENDING. The default value is DESCENDING.

sort_sortBy :: Lens' Sort (Maybe Text)

For ListEntities, supported attributes include LastModifiedDate (default), Visibility, EntityId, and Name.

For ListChangeSets, supported attributes include StartTime and EndTime.