Amazonka.MarketplaceCatalog.ListChangeSets

Description

Returns the list of change sets owned by the account being used to make the call. You can filter this list by providing any combination of entityId, ChangeSetName, and status. If you provide more than one filter, the API operation applies a logical AND between the filters.

You can describe a change during the 60-day request history retention period for API calls.

Creating a Request

data ListChangeSets

See: newListChangeSets smart constructor.

Constructors

ListChangeSets'

Fields

nextToken :: Maybe Text

The token value retrieved from a previous call to access the next page of results.

filterList :: Maybe (NonEmpty Filter)

An array of filter objects.

sort :: Maybe Sort

An object that contains two attributes, SortBy and SortOrder.

maxResults :: Maybe Natural

The maximum number of results returned by a single call. This value must be provided in the next call to retrieve the next set of results. By default, this value is 20.

catalog :: Text

The catalog related to the request. Fixed value: AWSMarketplace

newListChangeSets

Arguments

:: Text
$sel:catalog:ListChangeSets'

-> ListChangeSets

Create a value of ListChangeSets with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:nextToken:ListChangeSets', listChangeSets_nextToken - The token value retrieved from a previous call to access the next page of results.

$sel:filterList:ListChangeSets', listChangeSets_filterList - An array of filter objects.

$sel:sort:ListChangeSets', listChangeSets_sort - An object that contains two attributes, SortBy and SortOrder.

$sel:maxResults:ListChangeSets', listChangeSets_maxResults - The maximum number of results returned by a single call. This value must be provided in the next call to retrieve the next set of results. By default, this value is 20.

$sel:catalog:ListChangeSets', listChangeSets_catalog - The catalog related to the request. Fixed value: AWSMarketplace

Request Lenses

listChangeSets_nextToken :: Lens' ListChangeSets (Maybe Text)

The token value retrieved from a previous call to access the next page of results.

listChangeSets_filterList :: Lens' ListChangeSets (Maybe (NonEmpty Filter))

An array of filter objects.

listChangeSets_sort :: Lens' ListChangeSets (Maybe Sort)

An object that contains two attributes, SortBy and SortOrder.

listChangeSets_maxResults :: Lens' ListChangeSets (Maybe Natural)

The maximum number of results returned by a single call. This value must be provided in the next call to retrieve the next set of results. By default, this value is 20.

listChangeSets_catalog :: Lens' ListChangeSets Text

The catalog related to the request. Fixed value: AWSMarketplace

Destructuring the Response

data ListChangeSetsResponse

See: newListChangeSetsResponse smart constructor.

Constructors

ListChangeSetsResponse'

Fields

nextToken :: Maybe Text

The value of the next token, if it exists. Null if there are no more results.

changeSetSummaryList :: Maybe [ChangeSetSummaryListItem]

Array of ChangeSetSummaryListItem objects.

httpStatus :: Int

The response's http status code.

newListChangeSetsResponse

Arguments

:: Int
$sel:httpStatus:ListChangeSetsResponse'

-> ListChangeSetsResponse

Create a value of ListChangeSetsResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:nextToken:ListChangeSets', listChangeSetsResponse_nextToken - The value of the next token, if it exists. Null if there are no more results.

$sel:changeSetSummaryList:ListChangeSetsResponse', listChangeSetsResponse_changeSetSummaryList - Array of ChangeSetSummaryListItem objects.

$sel:httpStatus:ListChangeSetsResponse', listChangeSetsResponse_httpStatus - The response's http status code.

Response Lenses

listChangeSetsResponse_nextToken :: Lens' ListChangeSetsResponse (Maybe Text)

The value of the next token, if it exists. Null if there are no more results.

listChangeSetsResponse_changeSetSummaryList :: Lens' ListChangeSetsResponse (Maybe [ChangeSetSummaryListItem])

Array of ChangeSetSummaryListItem objects.

listChangeSetsResponse_httpStatus :: Lens' ListChangeSetsResponse Int

The response's http status code.