Amazonka.MarketplaceCatalog.StartChangeSet

Description

This operation allows you to request changes for your entities. Within a single ChangeSet, you cannot start the same change type against the same entity multiple times. Additionally, when a ChangeSet is running, all the entities targeted by the different changes are locked until the ChangeSet has completed (either succeeded, cancelled, or failed). If you try to start a ChangeSet containing a change against an entity that is already locked, you will receive a ResourceInUseException.

For example, you cannot start the ChangeSet described in the example later in this topic, because it contains two changes to execute the same change type (AddRevisions) against the same entity (entity-id@1).

For more information about working with change sets, see Working with change sets.

Creating a Request

data StartChangeSet

See: newStartChangeSet smart constructor.

Constructors

StartChangeSet'

Fields

changeSetName :: Maybe Text

Optional case sensitive string of up to 100 ASCII characters. The change set name can be used to filter the list of change sets.

clientRequestToken :: Maybe Text

A unique token to identify the request to ensure idempotency.

catalog :: Text

The catalog related to the request. Fixed value: AWSMarketplace

changeSet :: NonEmpty Change

Array of change object.

newStartChangeSet

Arguments

:: Text
$sel:catalog:StartChangeSet'

-> NonEmpty Change
$sel:changeSet:StartChangeSet'

-> StartChangeSet

Create a value of StartChangeSet with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:changeSetName:StartChangeSet', startChangeSet_changeSetName - Optional case sensitive string of up to 100 ASCII characters. The change set name can be used to filter the list of change sets.

$sel:clientRequestToken:StartChangeSet', startChangeSet_clientRequestToken - A unique token to identify the request to ensure idempotency.

$sel:catalog:StartChangeSet', startChangeSet_catalog - The catalog related to the request. Fixed value: AWSMarketplace

$sel:changeSet:StartChangeSet', startChangeSet_changeSet - Array of change object.

Request Lenses

startChangeSet_changeSetName :: Lens' StartChangeSet (Maybe Text)

Optional case sensitive string of up to 100 ASCII characters. The change set name can be used to filter the list of change sets.

startChangeSet_clientRequestToken :: Lens' StartChangeSet (Maybe Text)

A unique token to identify the request to ensure idempotency.

startChangeSet_catalog :: Lens' StartChangeSet Text

The catalog related to the request. Fixed value: AWSMarketplace

startChangeSet_changeSet :: Lens' StartChangeSet (NonEmpty Change)

Array of change object.

Destructuring the Response

data StartChangeSetResponse

See: newStartChangeSetResponse smart constructor.

Constructors

StartChangeSetResponse'

Fields

changeSetId :: Maybe Text

Unique identifier generated for the request.

changeSetArn :: Maybe Text

The ARN associated to the unique identifier generated for the request.

httpStatus :: Int

The response's http status code.

newStartChangeSetResponse

Arguments

:: Int
$sel:httpStatus:StartChangeSetResponse'

-> StartChangeSetResponse

Create a value of StartChangeSetResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:changeSetId:StartChangeSetResponse', startChangeSetResponse_changeSetId - Unique identifier generated for the request.

$sel:changeSetArn:StartChangeSetResponse', startChangeSetResponse_changeSetArn - The ARN associated to the unique identifier generated for the request.

$sel:httpStatus:StartChangeSetResponse', startChangeSetResponse_httpStatus - The response's http status code.

Response Lenses

startChangeSetResponse_changeSetId :: Lens' StartChangeSetResponse (Maybe Text)

Unique identifier generated for the request.

startChangeSetResponse_changeSetArn :: Lens' StartChangeSetResponse (Maybe Text)

The ARN associated to the unique identifier generated for the request.

startChangeSetResponse_httpStatus :: Lens' StartChangeSetResponse Int

The response's http status code.