Amazonka.MarketplaceCatalog.Types

Description

Service Configuration

defaultService :: Service

API version 2018-09-17 of the Amazon Marketplace Catalog Service SDK configuration.

Errors

_ValidationException :: AsError a => Getting (First ServiceError) a ServiceError

An error occurred during validation.

_AccessDeniedException :: AsError a => Getting (First ServiceError) a ServiceError

Access is denied.

_ServiceQuotaExceededException :: AsError a => Getting (First ServiceError) a ServiceError

The maximum number of open requests per account has been exceeded.

_ThrottlingException :: AsError a => Getting (First ServiceError) a ServiceError

Too many requests.

_InternalServiceException :: AsError a => Getting (First ServiceError) a ServiceError

There was an internal service exception.

_ResourceNotSupportedException :: AsError a => Getting (First ServiceError) a ServiceError

Currently, the specified resource is not supported.

_ResourceNotFoundException :: AsError a => Getting (First ServiceError) a ServiceError

The specified resource wasn't found.

_ResourceInUseException :: AsError a => Getting (First ServiceError) a ServiceError

The resource is currently in use.

ChangeStatus

newtype ChangeStatus

Constructors

ChangeStatus'

Fields

fromChangeStatus :: Text

Bundled Patterns

pattern ChangeStatus_APPLYING :: ChangeStatus

pattern ChangeStatus_CANCELLED :: ChangeStatus

pattern ChangeStatus_FAILED :: ChangeStatus

pattern ChangeStatus_PREPARING :: ChangeStatus

pattern ChangeStatus_SUCCEEDED :: ChangeStatus

FailureCode

newtype FailureCode

Constructors

FailureCode'

Fields

fromFailureCode :: Text

Bundled Patterns

pattern FailureCode_CLIENT_ERROR :: FailureCode

pattern FailureCode_SERVER_FAULT :: FailureCode

SortOrder

newtype SortOrder

Constructors

SortOrder'

Fields

fromSortOrder :: Text

Bundled Patterns

pattern SortOrder_ASCENDING :: SortOrder

pattern SortOrder_DESCENDING :: SortOrder

Change

data Change

An object that contains the ChangeType, Details, and Entity.

See: newChange smart constructor.

Constructors

Change'

Fields

changeName :: Maybe Text

Optional name for the change.

changeType :: Text

Change types are single string values that describe your intention for the change. Each change type is unique for each EntityType provided in the change's scope.

entity :: Entity

The entity to be changed.

details :: Text

This object contains details specific to the change type of the requested change.

newChange

Arguments

:: Text
$sel:changeType:Change'

-> Entity
$sel:entity:Change'

-> Text
$sel:details:Change'

-> Change

Create a value of Change with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:changeName:Change', change_changeName - Optional name for the change.

$sel:changeType:Change', change_changeType - Change types are single string values that describe your intention for the change. Each change type is unique for each EntityType provided in the change's scope.

$sel:entity:Change', change_entity - The entity to be changed.

$sel:details:Change', change_details - This object contains details specific to the change type of the requested change.

change_changeName :: Lens' Change (Maybe Text)

Optional name for the change.

change_changeType :: Lens' Change Text

Change types are single string values that describe your intention for the change. Each change type is unique for each EntityType provided in the change's scope.

change_entity :: Lens' Change Entity

The entity to be changed.

change_details :: Lens' Change Text

This object contains details specific to the change type of the requested change.

ChangeSetSummaryListItem

data ChangeSetSummaryListItem

A summary of a change set returned in a list of change sets when the ListChangeSets action is called.

See: newChangeSetSummaryListItem smart constructor.

Constructors

ChangeSetSummaryListItem'

Fields

status :: Maybe ChangeStatus

The current status of the change set.

entityIdList :: Maybe [Text]

This object is a list of entity IDs (string) that are a part of a change set. The entity ID list is a maximum of 20 entities. It must contain at least one entity.

startTime :: Maybe Text

The time, in ISO 8601 format (2018-02-27T13:45:22Z), when the change set was started.

failureCode :: Maybe FailureCode

Returned if the change set is in FAILED status. Can be either CLIENT_ERROR, which means that there are issues with the request (see the ErrorDetailList of DescribeChangeSet), or SERVER_FAULT, which means that there is a problem in the system, and you should retry your request.

changeSetName :: Maybe Text

The non-unique name for the change set.

changeSetId :: Maybe Text

The unique identifier for a change set.

endTime :: Maybe Text

The time, in ISO 8601 format (2018-02-27T13:45:22Z), when the change set was finished.

changeSetArn :: Maybe Text

The ARN associated with the unique identifier for the change set referenced in this request.

newChangeSetSummaryListItem :: ChangeSetSummaryListItem

Create a value of ChangeSetSummaryListItem with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:status:ChangeSetSummaryListItem', changeSetSummaryListItem_status - The current status of the change set.

$sel:entityIdList:ChangeSetSummaryListItem', changeSetSummaryListItem_entityIdList - This object is a list of entity IDs (string) that are a part of a change set. The entity ID list is a maximum of 20 entities. It must contain at least one entity.

$sel:startTime:ChangeSetSummaryListItem', changeSetSummaryListItem_startTime - The time, in ISO 8601 format (2018-02-27T13:45:22Z), when the change set was started.

$sel:failureCode:ChangeSetSummaryListItem', changeSetSummaryListItem_failureCode - Returned if the change set is in FAILED status. Can be either CLIENT_ERROR, which means that there are issues with the request (see the ErrorDetailList of DescribeChangeSet), or SERVER_FAULT, which means that there is a problem in the system, and you should retry your request.

$sel:changeSetName:ChangeSetSummaryListItem', changeSetSummaryListItem_changeSetName - The non-unique name for the change set.

$sel:changeSetId:ChangeSetSummaryListItem', changeSetSummaryListItem_changeSetId - The unique identifier for a change set.

$sel:endTime:ChangeSetSummaryListItem', changeSetSummaryListItem_endTime - The time, in ISO 8601 format (2018-02-27T13:45:22Z), when the change set was finished.

$sel:changeSetArn:ChangeSetSummaryListItem', changeSetSummaryListItem_changeSetArn - The ARN associated with the unique identifier for the change set referenced in this request.

changeSetSummaryListItem_status :: Lens' ChangeSetSummaryListItem (Maybe ChangeStatus)

The current status of the change set.

changeSetSummaryListItem_entityIdList :: Lens' ChangeSetSummaryListItem (Maybe [Text])

This object is a list of entity IDs (string) that are a part of a change set. The entity ID list is a maximum of 20 entities. It must contain at least one entity.

changeSetSummaryListItem_startTime :: Lens' ChangeSetSummaryListItem (Maybe Text)

The time, in ISO 8601 format (2018-02-27T13:45:22Z), when the change set was started.

changeSetSummaryListItem_failureCode :: Lens' ChangeSetSummaryListItem (Maybe FailureCode)

Returned if the change set is in FAILED status. Can be either CLIENT_ERROR, which means that there are issues with the request (see the ErrorDetailList of DescribeChangeSet), or SERVER_FAULT, which means that there is a problem in the system, and you should retry your request.

changeSetSummaryListItem_changeSetName :: Lens' ChangeSetSummaryListItem (Maybe Text)

The non-unique name for the change set.

changeSetSummaryListItem_changeSetId :: Lens' ChangeSetSummaryListItem (Maybe Text)

The unique identifier for a change set.

changeSetSummaryListItem_endTime :: Lens' ChangeSetSummaryListItem (Maybe Text)

The time, in ISO 8601 format (2018-02-27T13:45:22Z), when the change set was finished.

changeSetSummaryListItem_changeSetArn :: Lens' ChangeSetSummaryListItem (Maybe Text)

The ARN associated with the unique identifier for the change set referenced in this request.

ChangeSummary

data ChangeSummary

This object is a container for common summary information about the change. The summary doesn't contain the whole change structure.

See: newChangeSummary smart constructor.

Constructors

ChangeSummary'

Fields

changeName :: Maybe Text

Optional name for the change.

details :: Maybe Text

This object contains details specific to the change type of the requested change.

errorDetailList :: Maybe [ErrorDetail]

An array of ErrorDetail objects associated with the change.

entity :: Maybe Entity

The entity to be changed.

changeType :: Maybe Text

The type of the change.

newChangeSummary :: ChangeSummary

Create a value of ChangeSummary with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:changeName:ChangeSummary', changeSummary_changeName - Optional name for the change.

$sel:details:ChangeSummary', changeSummary_details - This object contains details specific to the change type of the requested change.

$sel:errorDetailList:ChangeSummary', changeSummary_errorDetailList - An array of ErrorDetail objects associated with the change.

$sel:entity:ChangeSummary', changeSummary_entity - The entity to be changed.

$sel:changeType:ChangeSummary', changeSummary_changeType - The type of the change.

changeSummary_changeName :: Lens' ChangeSummary (Maybe Text)

Optional name for the change.

changeSummary_details :: Lens' ChangeSummary (Maybe Text)

This object contains details specific to the change type of the requested change.

changeSummary_errorDetailList :: Lens' ChangeSummary (Maybe [ErrorDetail])

An array of ErrorDetail objects associated with the change.

changeSummary_entity :: Lens' ChangeSummary (Maybe Entity)

The entity to be changed.

changeSummary_changeType :: Lens' ChangeSummary (Maybe Text)

The type of the change.

Entity

data Entity

An entity contains data that describes your product, its supported features, and how it can be used or launched by your customer.

See: newEntity smart constructor.

Constructors

Entity'

Fields

identifier :: Maybe Text

The identifier for the entity.

type' :: Text

The type of entity.

newEntity

Arguments

:: Text
$sel:type':Entity'

-> Entity

Create a value of Entity with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:identifier:Entity', entity_identifier - The identifier for the entity.

$sel:type':Entity', entity_type - The type of entity.

entity_identifier :: Lens' Entity (Maybe Text)

The identifier for the entity.

entity_type :: Lens' Entity Text

The type of entity.

EntitySummary

data EntitySummary

This object is a container for common summary information about the entity. The summary doesn't contain the whole entity structure, but it does contain information common across all entities.

See: newEntitySummary smart constructor.

Constructors

EntitySummary'

Fields

lastModifiedDate :: Maybe Text

The last time the entity was published, using ISO 8601 format (2018-02-27T13:45:22Z).

entityType :: Maybe Text

The type of the entity.

visibility :: Maybe Text

The visibility status of the entity to buyers. This value can be Public (everyone can view the entity), Limited (the entity is visible to limited accounts only), or Restricted (the entity was published and then unpublished and only existing buyers can view it).

name :: Maybe Text

The name for the entity. This value is not unique. It is defined by the seller.

entityId :: Maybe Text

The unique identifier for the entity.

entityArn :: Maybe Text

The ARN associated with the unique identifier for the entity.

newEntitySummary :: EntitySummary

Create a value of EntitySummary with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:lastModifiedDate:EntitySummary', entitySummary_lastModifiedDate - The last time the entity was published, using ISO 8601 format (2018-02-27T13:45:22Z).

$sel:entityType:EntitySummary', entitySummary_entityType - The type of the entity.

$sel:visibility:EntitySummary', entitySummary_visibility - The visibility status of the entity to buyers. This value can be Public (everyone can view the entity), Limited (the entity is visible to limited accounts only), or Restricted (the entity was published and then unpublished and only existing buyers can view it).

$sel:name:EntitySummary', entitySummary_name - The name for the entity. This value is not unique. It is defined by the seller.

$sel:entityId:EntitySummary', entitySummary_entityId - The unique identifier for the entity.

$sel:entityArn:EntitySummary', entitySummary_entityArn - The ARN associated with the unique identifier for the entity.

entitySummary_lastModifiedDate :: Lens' EntitySummary (Maybe Text)

The last time the entity was published, using ISO 8601 format (2018-02-27T13:45:22Z).

entitySummary_entityType :: Lens' EntitySummary (Maybe Text)

The type of the entity.

entitySummary_visibility :: Lens' EntitySummary (Maybe Text)

The visibility status of the entity to buyers. This value can be Public (everyone can view the entity), Limited (the entity is visible to limited accounts only), or Restricted (the entity was published and then unpublished and only existing buyers can view it).

entitySummary_name :: Lens' EntitySummary (Maybe Text)

The name for the entity. This value is not unique. It is defined by the seller.

entitySummary_entityId :: Lens' EntitySummary (Maybe Text)

The unique identifier for the entity.

entitySummary_entityArn :: Lens' EntitySummary (Maybe Text)

The ARN associated with the unique identifier for the entity.

ErrorDetail

data ErrorDetail

Details about the error.

See: newErrorDetail smart constructor.

Constructors

ErrorDetail'

Fields

errorCode :: Maybe Text

The error code that identifies the type of error.

errorMessage :: Maybe Text

The message for the error.

newErrorDetail :: ErrorDetail

Create a value of ErrorDetail with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:errorCode:ErrorDetail', errorDetail_errorCode - The error code that identifies the type of error.

$sel:errorMessage:ErrorDetail', errorDetail_errorMessage - The message for the error.

errorDetail_errorCode :: Lens' ErrorDetail (Maybe Text)

The error code that identifies the type of error.

errorDetail_errorMessage :: Lens' ErrorDetail (Maybe Text)

The message for the error.

Filter

data Filter

A filter object, used to optionally filter results from calls to the ListEntities and ListChangeSets actions.

See: newFilter smart constructor.

Constructors

Filter'

Fields

valueList :: Maybe (NonEmpty Text)

ListEntities - This is a list of unique EntityIds.

ListChangeSets - The supported filter names and associated ValueLists is as follows:

ChangeSetName - The supported ValueList is a list of non-unique ChangeSetNames. These are defined when you call the StartChangeSet action.

Status - The supported ValueList is a list of statuses for all change set requests.

EntityId - The supported ValueList is a list of unique EntityIds.

BeforeStartTime - The supported ValueList is a list of all change sets that started before the filter value.

AfterStartTime - The supported ValueList is a list of all change sets that started after the filter value.

BeforeEndTime - The supported ValueList is a list of all change sets that ended before the filter value.

AfterEndTime - The supported ValueList is a list of all change sets that ended after the filter value.

name :: Maybe Text

For ListEntities, the supported value for this is an EntityId.

For ListChangeSets, the supported values are as follows:

newFilter :: Filter

Create a value of Filter with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:valueList:Filter', filter_valueList - ListEntities - This is a list of unique EntityIds.

ListChangeSets - The supported filter names and associated ValueLists is as follows:

ChangeSetName - The supported ValueList is a list of non-unique ChangeSetNames. These are defined when you call the StartChangeSet action.

Status - The supported ValueList is a list of statuses for all change set requests.

EntityId - The supported ValueList is a list of unique EntityIds.

BeforeStartTime - The supported ValueList is a list of all change sets that started before the filter value.

AfterStartTime - The supported ValueList is a list of all change sets that started after the filter value.

BeforeEndTime - The supported ValueList is a list of all change sets that ended before the filter value.

AfterEndTime - The supported ValueList is a list of all change sets that ended after the filter value.

$sel:name:Filter', filter_name - For ListEntities, the supported value for this is an EntityId.

For ListChangeSets, the supported values are as follows:

filter_valueList :: Lens' Filter (Maybe (NonEmpty Text))

ListEntities - This is a list of unique EntityIds.

ListChangeSets - The supported filter names and associated ValueLists is as follows:

ChangeSetName - The supported ValueList is a list of non-unique ChangeSetNames. These are defined when you call the StartChangeSet action.

Status - The supported ValueList is a list of statuses for all change set requests.

EntityId - The supported ValueList is a list of unique EntityIds.

BeforeStartTime - The supported ValueList is a list of all change sets that started before the filter value.

AfterStartTime - The supported ValueList is a list of all change sets that started after the filter value.

BeforeEndTime - The supported ValueList is a list of all change sets that ended before the filter value.

AfterEndTime - The supported ValueList is a list of all change sets that ended after the filter value.

filter_name :: Lens' Filter (Maybe Text)

For ListEntities, the supported value for this is an EntityId.

For ListChangeSets, the supported values are as follows:

Sort

data Sort

An object that contains two attributes, SortBy and SortOrder.

See: newSort smart constructor.

Constructors

Sort'

Fields

sortOrder :: Maybe SortOrder

The sorting order. Can be ASCENDING or DESCENDING. The default value is DESCENDING.

sortBy :: Maybe Text

For ListEntities, supported attributes include LastModifiedDate (default), Visibility, EntityId, and Name.

For ListChangeSets, supported attributes include StartTime and EndTime.

newSort :: Sort

Create a value of Sort with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:sortOrder:Sort', sort_sortOrder - The sorting order. Can be ASCENDING or DESCENDING. The default value is DESCENDING.

$sel:sortBy:Sort', sort_sortBy - For ListEntities, supported attributes include LastModifiedDate (default), Visibility, EntityId, and Name.

For ListChangeSets, supported attributes include StartTime and EndTime.

sort_sortOrder :: Lens' Sort (Maybe SortOrder)

The sorting order. Can be ASCENDING or DESCENDING. The default value is DESCENDING.

sort_sortBy :: Lens' Sort (Maybe Text)

For ListEntities, supported attributes include LastModifiedDate (default), Visibility, EntityId, and Name.

For ListChangeSets, supported attributes include StartTime and EndTime.