Amazonka.MarketplaceCatalog.Types.ChangeSetSummaryListItem

Description

Documentation

data ChangeSetSummaryListItem

A summary of a change set returned in a list of change sets when the ListChangeSets action is called.

See: newChangeSetSummaryListItem smart constructor.

Constructors

ChangeSetSummaryListItem'

Fields

status :: Maybe ChangeStatus

The current status of the change set.

entityIdList :: Maybe [Text]

This object is a list of entity IDs (string) that are a part of a change set. The entity ID list is a maximum of 20 entities. It must contain at least one entity.

startTime :: Maybe Text

The time, in ISO 8601 format (2018-02-27T13:45:22Z), when the change set was started.

failureCode :: Maybe FailureCode

Returned if the change set is in FAILED status. Can be either CLIENT_ERROR, which means that there are issues with the request (see the ErrorDetailList of DescribeChangeSet), or SERVER_FAULT, which means that there is a problem in the system, and you should retry your request.

changeSetName :: Maybe Text

The non-unique name for the change set.

changeSetId :: Maybe Text

The unique identifier for a change set.

endTime :: Maybe Text

The time, in ISO 8601 format (2018-02-27T13:45:22Z), when the change set was finished.

changeSetArn :: Maybe Text

The ARN associated with the unique identifier for the change set referenced in this request.

newChangeSetSummaryListItem :: ChangeSetSummaryListItem

Create a value of ChangeSetSummaryListItem with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:status:ChangeSetSummaryListItem', changeSetSummaryListItem_status - The current status of the change set.

$sel:entityIdList:ChangeSetSummaryListItem', changeSetSummaryListItem_entityIdList - This object is a list of entity IDs (string) that are a part of a change set. The entity ID list is a maximum of 20 entities. It must contain at least one entity.

$sel:startTime:ChangeSetSummaryListItem', changeSetSummaryListItem_startTime - The time, in ISO 8601 format (2018-02-27T13:45:22Z), when the change set was started.

$sel:failureCode:ChangeSetSummaryListItem', changeSetSummaryListItem_failureCode - Returned if the change set is in FAILED status. Can be either CLIENT_ERROR, which means that there are issues with the request (see the ErrorDetailList of DescribeChangeSet), or SERVER_FAULT, which means that there is a problem in the system, and you should retry your request.

$sel:changeSetName:ChangeSetSummaryListItem', changeSetSummaryListItem_changeSetName - The non-unique name for the change set.

$sel:changeSetId:ChangeSetSummaryListItem', changeSetSummaryListItem_changeSetId - The unique identifier for a change set.

$sel:endTime:ChangeSetSummaryListItem', changeSetSummaryListItem_endTime - The time, in ISO 8601 format (2018-02-27T13:45:22Z), when the change set was finished.

$sel:changeSetArn:ChangeSetSummaryListItem', changeSetSummaryListItem_changeSetArn - The ARN associated with the unique identifier for the change set referenced in this request.

changeSetSummaryListItem_status :: Lens' ChangeSetSummaryListItem (Maybe ChangeStatus)

The current status of the change set.

changeSetSummaryListItem_entityIdList :: Lens' ChangeSetSummaryListItem (Maybe [Text])

This object is a list of entity IDs (string) that are a part of a change set. The entity ID list is a maximum of 20 entities. It must contain at least one entity.

changeSetSummaryListItem_startTime :: Lens' ChangeSetSummaryListItem (Maybe Text)

The time, in ISO 8601 format (2018-02-27T13:45:22Z), when the change set was started.

changeSetSummaryListItem_failureCode :: Lens' ChangeSetSummaryListItem (Maybe FailureCode)

Returned if the change set is in FAILED status. Can be either CLIENT_ERROR, which means that there are issues with the request (see the ErrorDetailList of DescribeChangeSet), or SERVER_FAULT, which means that there is a problem in the system, and you should retry your request.

changeSetSummaryListItem_changeSetName :: Lens' ChangeSetSummaryListItem (Maybe Text)

The non-unique name for the change set.

changeSetSummaryListItem_changeSetId :: Lens' ChangeSetSummaryListItem (Maybe Text)

The unique identifier for a change set.

changeSetSummaryListItem_endTime :: Lens' ChangeSetSummaryListItem (Maybe Text)

The time, in ISO 8601 format (2018-02-27T13:45:22Z), when the change set was finished.

changeSetSummaryListItem_changeSetArn :: Lens' ChangeSetSummaryListItem (Maybe Text)

The ARN associated with the unique identifier for the change set referenced in this request.