Amazonka.MarketplaceEntitlement

Description

Derived from API version 2017-01-11 of the AWS service descriptions, licensed under Apache 2.0.

AWS Marketplace Entitlement Service

This reference provides descriptions of the AWS Marketplace Entitlement Service API.

AWS Marketplace Entitlement Service is used to determine the entitlement of a customer to a given product. An entitlement represents capacity in a product owned by the customer. For example, a customer might own some number of users or seats in an SaaS application or some amount of data capacity in a multi-tenant database.

Getting Entitlement Records

GetEntitlements- Gets the entitlements for a Marketplace product.

Service Configuration

defaultService :: Service

API version 2017-01-11 of the Amazon Marketplace Entitlement Service SDK configuration.

Errors

Error matchers are designed for use with the functions provided by Control.Exception.Lens. This allows catching (and rethrowing) service specific errors returned by MarketplaceEntitlement.

InvalidParameterException

_InvalidParameterException :: AsError a => Getting (First ServiceError) a ServiceError

One or more parameters in your request was invalid.

ThrottlingException

_ThrottlingException :: AsError a => Getting (First ServiceError) a ServiceError

The calls to the GetEntitlements API are throttled.

InternalServiceErrorException

_InternalServiceErrorException :: AsError a => Getting (First ServiceError) a ServiceError

An internal error has occurred. Retry your request. If the problem persists, post a message with details on the AWS forums.

Waiters

Waiters poll by repeatedly sending a request until some remote success condition configured by the Wait specification is fulfilled. The Wait specification determines how many attempts should be made, in addition to delay and retry strategies.

Operations

Some AWS operations return results that are incomplete and require subsequent requests in order to obtain the entire result set. The process of sending subsequent requests to continue where a previous request left off is called pagination. For example, the ListObjects operation of Amazon S3 returns up to 1000 objects at a time, and you must send subsequent requests with the appropriate Marker in order to retrieve the next page of results.

Operations that have an AWSPager instance can transparently perform subsequent requests, correctly setting Markers and other request facets to iterate through the entire result set of a truncated API operation. Operations which support this have an additional note in the documentation.

Many operations have the ability to filter results on the server side. See the individual operation parameters for details.

GetEntitlements (Paginated)

data GetEntitlements

The GetEntitlementsRequest contains parameters for the GetEntitlements operation.

See: newGetEntitlements smart constructor.

Constructors

GetEntitlements' (Maybe Text) (Maybe (HashMap GetEntitlementFilterName (NonEmpty Text))) (Maybe Int) Text

newGetEntitlements

Arguments

:: Text
$sel:productCode:GetEntitlements'

-> GetEntitlements

Create a value of GetEntitlements with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:nextToken:GetEntitlements', getEntitlements_nextToken - For paginated calls to GetEntitlements, pass the NextToken from the previous GetEntitlementsResult.

$sel:filter':GetEntitlements', getEntitlements_filter - Filter is used to return entitlements for a specific customer or for a specific dimension. Filters are described as keys mapped to a lists of values. Filtered requests are unioned for each value in the value list, and then intersected for each filter key.

$sel:maxResults:GetEntitlements', getEntitlements_maxResults - The maximum number of items to retrieve from the GetEntitlements operation. For pagination, use the NextToken field in subsequent calls to GetEntitlements.

$sel:productCode:GetEntitlements', getEntitlements_productCode - Product code is used to uniquely identify a product in AWS Marketplace. The product code will be provided by AWS Marketplace when the product listing is created.

data GetEntitlementsResponse

The GetEntitlementsRequest contains results from the GetEntitlements operation.

See: newGetEntitlementsResponse smart constructor.

Constructors

GetEntitlementsResponse' (Maybe Text) (Maybe [Entitlement]) Int

newGetEntitlementsResponse

Arguments

:: Int
$sel:httpStatus:GetEntitlementsResponse'

-> GetEntitlementsResponse

Create a value of GetEntitlementsResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:nextToken:GetEntitlements', getEntitlementsResponse_nextToken - For paginated results, use NextToken in subsequent calls to GetEntitlements. If the result contains an empty set of entitlements, NextToken might still be present and should be used.

$sel:entitlements:GetEntitlementsResponse', getEntitlementsResponse_entitlements - The set of entitlements found through the GetEntitlements operation. If the result contains an empty set of entitlements, NextToken might still be present and should be used.

$sel:httpStatus:GetEntitlementsResponse', getEntitlementsResponse_httpStatus - The response's http status code.

Types

GetEntitlementFilterName

newtype GetEntitlementFilterName

Constructors

GetEntitlementFilterName'

Fields

fromGetEntitlementFilterName :: Text

Bundled Patterns

pattern GetEntitlementFilterName_CUSTOMER_IDENTIFIER :: GetEntitlementFilterName

pattern GetEntitlementFilterName_DIMENSION :: GetEntitlementFilterName

Entitlement

data Entitlement

An entitlement represents capacity in a product owned by the customer. For example, a customer might own some number of users or seats in an SaaS application or some amount of data capacity in a multi-tenant database.

See: newEntitlement smart constructor.

Constructors

Entitlement' (Maybe Text) (Maybe EntitlementValue) (Maybe POSIX) (Maybe Text) (Maybe Text)

newEntitlement :: Entitlement

Create a value of Entitlement with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:dimension:Entitlement', entitlement_dimension - The dimension for which the given entitlement applies. Dimensions represent categories of capacity in a product and are specified when the product is listed in AWS Marketplace.

$sel:value:Entitlement', entitlement_value - The EntitlementValue represents the amount of capacity that the customer is entitled to for the product.

$sel:expirationDate:Entitlement', entitlement_expirationDate - The expiration date represents the minimum date through which this entitlement is expected to remain valid. For contractual products listed on AWS Marketplace, the expiration date is the date at which the customer will renew or cancel their contract. Customers who are opting to renew their contract will still have entitlements with an expiration date.

$sel:customerIdentifier:Entitlement', entitlement_customerIdentifier - The customer identifier is a handle to each unique customer in an application. Customer identifiers are obtained through the ResolveCustomer operation in AWS Marketplace Metering Service.

$sel:productCode:Entitlement', entitlement_productCode - The product code for which the given entitlement applies. Product codes are provided by AWS Marketplace when the product listing is created.

EntitlementValue

data EntitlementValue

The EntitlementValue represents the amount of capacity that the customer is entitled to for the product.

See: newEntitlementValue smart constructor.

Constructors

EntitlementValue' (Maybe Int) (Maybe Double) (Maybe Text) (Maybe Bool)

newEntitlementValue :: EntitlementValue

Create a value of EntitlementValue with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:integerValue:EntitlementValue', entitlementValue_integerValue - The IntegerValue field will be populated with an integer value when the entitlement is an integer type. Otherwise, the field will not be set.

$sel:doubleValue:EntitlementValue', entitlementValue_doubleValue - The DoubleValue field will be populated with a double value when the entitlement is a double type. Otherwise, the field will not be set.

$sel:stringValue:EntitlementValue', entitlementValue_stringValue - The StringValue field will be populated with a string value when the entitlement is a string type. Otherwise, the field will not be set.

$sel:booleanValue:EntitlementValue', entitlementValue_booleanValue - The BooleanValue field will be populated with a boolean value when the entitlement is a boolean type. Otherwise, the field will not be set.