Amazonka.MarketplaceMetering

Description

Derived from API version 2016-01-14 of the AWS service descriptions, licensed under Apache 2.0.

AWS Marketplace Metering Service

This reference provides descriptions of the low-level AWS Marketplace Metering Service API.

AWS Marketplace sellers can use this API to submit usage data for custom usage dimensions.

For information on the permissions you need to use this API, see AWS Marketing metering and entitlement API permissions in the AWS Marketplace Seller Guide.

Submitting Metering Records

MeterUsage- Submits the metering record for a Marketplace product. MeterUsage is called from an EC2 instance or a container running on EKS or ECS.

BatchMeterUsage- Submits the metering record for a set of customers. BatchMeterUsage is called from a software-as-a-service (SaaS) application.

Accepting New Customers

ResolveCustomer- Called by a SaaS application during the registration process. When a buyer visits your website during the registration process, the buyer submits a Registration Token through the browser. The Registration Token is resolved through this API to obtain a CustomerIdentifier and Product Code.

Entitlement and Metering for Paid Container Products

Paid container software products sold through AWS Marketplace must integrate with the AWS Marketplace Metering Service and call the RegisterUsage operation for software entitlement and metering. Free and BYOL products for Amazon ECS or Amazon EKS aren't required to call RegisterUsage, but you can do so if you want to receive usage data in your seller reports. For more information on using the RegisterUsage operation, see Container-Based Products.

BatchMeterUsage API calls are captured by AWS CloudTrail. You can use Cloudtrail to verify that the SaaS metering records that you sent are accurate by searching for records with the eventName of BatchMeterUsage. You can also use CloudTrail to audit records over time. For more information, see the <http:docs.aws.amazon.comawscloudtraillatestuserguidecloudtrail-concepts.html AWS CloudTrail User Guide> .

Service Configuration

defaultService :: Service

API version 2016-01-14 of the Amazon Marketplace Metering SDK configuration.

Errors

Error matchers are designed for use with the functions provided by Control.Exception.Lens. This allows catching (and rethrowing) service specific errors returned by MarketplaceMetering.

InvalidTagException

_InvalidTagException :: AsError a => Getting (First ServiceError) a ServiceError

The tag is invalid, or the number of tags is greater than 5.

InvalidEndpointRegionException

_InvalidEndpointRegionException :: AsError a => Getting (First ServiceError) a ServiceError

The endpoint being called is in a AWS Region different from your EC2 instance, ECS task, or EKS pod. The Region of the Metering Service endpoint and the AWS Region of the resource must match.

InvalidProductCodeException

_InvalidProductCodeException :: AsError a => Getting (First ServiceError) a ServiceError

The product code passed does not match the product code used for publishing the product.

InvalidUsageDimensionException

_InvalidUsageDimensionException :: AsError a => Getting (First ServiceError) a ServiceError

The usage dimension does not match one of the UsageDimensions associated with products.

PlatformNotSupportedException

_PlatformNotSupportedException :: AsError a => Getting (First ServiceError) a ServiceError

AWS Marketplace does not support metering usage from the underlying platform. Currently, Amazon ECS, Amazon EKS, and AWS Fargate are supported.

CustomerNotEntitledException

_CustomerNotEntitledException :: AsError a => Getting (First ServiceError) a ServiceError

Exception thrown when the customer does not have a valid subscription for the product.

DuplicateRequestException

_DuplicateRequestException :: AsError a => Getting (First ServiceError) a ServiceError

A metering record has already been emitted by the same EC2 instance, ECS task, or EKS pod for the given {usageDimension, timestamp} with a different usageQuantity.

DisabledApiException

_DisabledApiException :: AsError a => Getting (First ServiceError) a ServiceError

The API is disabled in the Region.

TimestampOutOfBoundsException

_TimestampOutOfBoundsException :: AsError a => Getting (First ServiceError) a ServiceError

The timestamp value passed in the meterUsage() is out of allowed range.

ThrottlingException

_ThrottlingException :: AsError a => Getting (First ServiceError) a ServiceError

The calls to the API are throttled.

InvalidPublicKeyVersionException

_InvalidPublicKeyVersionException :: AsError a => Getting (First ServiceError) a ServiceError

Public Key version is invalid.

InternalServiceErrorException

_InternalServiceErrorException :: AsError a => Getting (First ServiceError) a ServiceError

An internal error has occurred. Retry your request. If the problem persists, post a message with details on the AWS forums.

InvalidTokenException

_InvalidTokenException :: AsError a => Getting (First ServiceError) a ServiceError

Registration token is invalid.

InvalidUsageAllocationsException

_InvalidUsageAllocationsException :: AsError a => Getting (First ServiceError) a ServiceError

The usage allocation objects are invalid, or the number of allocations is greater than 500 for a single usage record.

ExpiredTokenException

_ExpiredTokenException :: AsError a => Getting (First ServiceError) a ServiceError

The submitted registration token has expired. This can happen if the buyer's browser takes too long to redirect to your page, the buyer has resubmitted the registration token, or your application has held on to the registration token for too long. Your SaaS registration website should redeem this token as soon as it is submitted by the buyer's browser.

InvalidRegionException

_InvalidRegionException :: AsError a => Getting (First ServiceError) a ServiceError

RegisterUsage must be called in the same AWS Region the ECS task was launched in. This prevents a container from hardcoding a Region (e.g. withRegion(“us-east-1”) when calling RegisterUsage.

InvalidCustomerIdentifierException

_InvalidCustomerIdentifierException :: AsError a => Getting (First ServiceError) a ServiceError

You have metered usage for a CustomerIdentifier that does not exist.

Waiters

Waiters poll by repeatedly sending a request until some remote success condition configured by the Wait specification is fulfilled. The Wait specification determines how many attempts should be made, in addition to delay and retry strategies.

Operations

Some AWS operations return results that are incomplete and require subsequent requests in order to obtain the entire result set. The process of sending subsequent requests to continue where a previous request left off is called pagination. For example, the ListObjects operation of Amazon S3 returns up to 1000 objects at a time, and you must send subsequent requests with the appropriate Marker in order to retrieve the next page of results.

Operations that have an AWSPager instance can transparently perform subsequent requests, correctly setting Markers and other request facets to iterate through the entire result set of a truncated API operation. Operations which support this have an additional note in the documentation.

Many operations have the ability to filter results on the server side. See the individual operation parameters for details.

RegisterUsage

data RegisterUsage

See: newRegisterUsage smart constructor.

Constructors

RegisterUsage' (Maybe Text) Text Natural

newRegisterUsage

Arguments

:: Text
$sel:productCode:RegisterUsage'

-> Natural
$sel:publicKeyVersion:RegisterUsage'

-> RegisterUsage

Create a value of RegisterUsage with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:nonce:RegisterUsage', registerUsage_nonce - (Optional) To scope down the registration to a specific running software instance and guard against replay attacks.

$sel:productCode:RegisterUsage', registerUsage_productCode - Product code is used to uniquely identify a product in AWS Marketplace. The product code should be the same as the one used during the publishing of a new product.

$sel:publicKeyVersion:RegisterUsage', registerUsage_publicKeyVersion - Public Key Version provided by AWS Marketplace

data RegisterUsageResponse

See: newRegisterUsageResponse smart constructor.

Constructors

RegisterUsageResponse' (Maybe Text) (Maybe POSIX) Int

newRegisterUsageResponse

Arguments

:: Int
$sel:httpStatus:RegisterUsageResponse'

-> RegisterUsageResponse

Create a value of RegisterUsageResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:signature:RegisterUsageResponse', registerUsageResponse_signature - JWT Token

$sel:publicKeyRotationTimestamp:RegisterUsageResponse', registerUsageResponse_publicKeyRotationTimestamp - (Optional) Only included when public key version has expired

$sel:httpStatus:RegisterUsageResponse', registerUsageResponse_httpStatus - The response's http status code.

BatchMeterUsage

data BatchMeterUsage

A BatchMeterUsageRequest contains UsageRecords, which indicate quantities of usage within your application.

See: newBatchMeterUsage smart constructor.

Constructors

BatchMeterUsage' [UsageRecord] Text

newBatchMeterUsage

Arguments

:: Text
$sel:productCode:BatchMeterUsage'

-> BatchMeterUsage

Create a value of BatchMeterUsage with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:usageRecords:BatchMeterUsage', batchMeterUsage_usageRecords - The set of UsageRecords to submit. BatchMeterUsage accepts up to 25 UsageRecords at a time.

$sel:productCode:BatchMeterUsage', batchMeterUsage_productCode - Product code is used to uniquely identify a product in AWS Marketplace. The product code should be the same as the one used during the publishing of a new product.

data BatchMeterUsageResponse

Contains the UsageRecords processed by BatchMeterUsage and any records that have failed due to transient error.

See: newBatchMeterUsageResponse smart constructor.

Constructors

BatchMeterUsageResponse' (Maybe [UsageRecordResult]) (Maybe [UsageRecord]) Int

newBatchMeterUsageResponse

Arguments

:: Int
$sel:httpStatus:BatchMeterUsageResponse'

-> BatchMeterUsageResponse

Create a value of BatchMeterUsageResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:results:BatchMeterUsageResponse', batchMeterUsageResponse_results - Contains all UsageRecords processed by BatchMeterUsage. These records were either honored by AWS Marketplace Metering Service or were invalid.

$sel:unprocessedRecords:BatchMeterUsageResponse', batchMeterUsageResponse_unprocessedRecords - Contains all UsageRecords that were not processed by BatchMeterUsage. This is a list of UsageRecords. You can retry the failed request by making another BatchMeterUsage call with this list as input in the BatchMeterUsageRequest.

$sel:httpStatus:BatchMeterUsageResponse', batchMeterUsageResponse_httpStatus - The response's http status code.

ResolveCustomer

data ResolveCustomer

Contains input to the ResolveCustomer operation.

See: newResolveCustomer smart constructor.

Constructors

ResolveCustomer' Text

newResolveCustomer

Arguments

:: Text
$sel:registrationToken:ResolveCustomer'

-> ResolveCustomer

Create a value of ResolveCustomer with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:registrationToken:ResolveCustomer', resolveCustomer_registrationToken - When a buyer visits your website during the registration process, the buyer submits a registration token through the browser. The registration token is resolved to obtain a CustomerIdentifier and product code.

data ResolveCustomerResponse

The result of the ResolveCustomer operation. Contains the CustomerIdentifier and product code.

See: newResolveCustomerResponse smart constructor.

Constructors

ResolveCustomerResponse' (Maybe Text) (Maybe Text) Int

newResolveCustomerResponse

Arguments

:: Int
$sel:httpStatus:ResolveCustomerResponse'

-> ResolveCustomerResponse

Create a value of ResolveCustomerResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:customerIdentifier:ResolveCustomerResponse', resolveCustomerResponse_customerIdentifier - The CustomerIdentifier is used to identify an individual customer in your application. Calls to BatchMeterUsage require CustomerIdentifiers for each UsageRecord.

$sel:productCode:ResolveCustomerResponse', resolveCustomerResponse_productCode - The product code is returned to confirm that the buyer is registering for your product. Subsequent BatchMeterUsage calls should be made using this product code.

$sel:httpStatus:ResolveCustomerResponse', resolveCustomerResponse_httpStatus - The response's http status code.

MeterUsage

data MeterUsage

See: newMeterUsage smart constructor.

Constructors

MeterUsage' (Maybe Natural) (Maybe (NonEmpty UsageAllocation)) (Maybe Bool) Text POSIX Text

newMeterUsage

Arguments

:: Text
$sel:productCode:MeterUsage'

-> UTCTime
$sel:timestamp:MeterUsage'

-> Text
$sel:usageDimension:MeterUsage'

-> MeterUsage

Create a value of MeterUsage with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:usageQuantity:MeterUsage', meterUsage_usageQuantity - Consumption value for the hour. Defaults to 0 if not specified.

$sel:usageAllocations:MeterUsage', meterUsage_usageAllocations - The set of UsageAllocations to submit.

The sum of all UsageAllocation quantities must equal the UsageQuantity of the MeterUsage request, and each UsageAllocation must have a unique set of tags (include no tags).

$sel:dryRun:MeterUsage', meterUsage_dryRun - Checks whether you have the permissions required for the action, but does not make the request. If you have the permissions, the request returns DryRunOperation; otherwise, it returns UnauthorizedException. Defaults to false if not specified.

$sel:productCode:MeterUsage', meterUsage_productCode - Product code is used to uniquely identify a product in AWS Marketplace. The product code should be the same as the one used during the publishing of a new product.

$sel:timestamp:MeterUsage', meterUsage_timestamp - Timestamp, in UTC, for which the usage is being reported. Your application can meter usage for up to one hour in the past. Make sure the timestamp value is not before the start of the software usage.

$sel:usageDimension:MeterUsage', meterUsage_usageDimension - It will be one of the fcp dimension name provided during the publishing of the product.

data MeterUsageResponse

See: newMeterUsageResponse smart constructor.

Constructors

MeterUsageResponse' (Maybe Text) Int

newMeterUsageResponse

Arguments

:: Int
$sel:httpStatus:MeterUsageResponse'

-> MeterUsageResponse

Create a value of MeterUsageResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:meteringRecordId:MeterUsageResponse', meterUsageResponse_meteringRecordId - Metering record id.

$sel:httpStatus:MeterUsageResponse', meterUsageResponse_httpStatus - The response's http status code.

Types

UsageRecordResultStatus

newtype UsageRecordResultStatus

Constructors

UsageRecordResultStatus'

Fields

fromUsageRecordResultStatus :: Text

Bundled Patterns

pattern UsageRecordResultStatus_CustomerNotSubscribed :: UsageRecordResultStatus

pattern UsageRecordResultStatus_DuplicateRecord :: UsageRecordResultStatus

pattern UsageRecordResultStatus_Success :: UsageRecordResultStatus

Tag

data Tag

Metadata assigned to an allocation. Each tag is made up of a key and a value.

See: newTag smart constructor.

Constructors

Tag' Text Text

newTag

Arguments

:: Text
$sel:key:Tag'

-> Text
$sel:value:Tag'

-> Tag

Create a value of Tag with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:key:Tag', tag_key - One part of a key-value pair that makes up a tag. A key is a label that acts like a category for the specific tag values.

$sel:value:Tag', tag_value - One part of a key-value pair that makes up a tag. A value acts as a descriptor within a tag category (key). The value can be empty or null.

UsageAllocation

data UsageAllocation

Usage allocations allow you to split usage into buckets by tags.

Each UsageAllocation indicates the usage quantity for a specific set of tags.

See: newUsageAllocation smart constructor.

Constructors

UsageAllocation' (Maybe (NonEmpty Tag)) Natural

newUsageAllocation

Arguments

:: Natural
$sel:allocatedUsageQuantity:UsageAllocation'

-> UsageAllocation

Create a value of UsageAllocation with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:tags:UsageAllocation', usageAllocation_tags - The set of tags that define the bucket of usage. For the bucket of items with no tags, this parameter can be left out.

$sel:allocatedUsageQuantity:UsageAllocation', usageAllocation_allocatedUsageQuantity - The total quantity allocated to this bucket of usage.

UsageRecord

data UsageRecord

A UsageRecord indicates a quantity of usage for a given product, customer, dimension and time.

Multiple requests with the same UsageRecords as input will be deduplicated to prevent double charges.

See: newUsageRecord smart constructor.

Constructors

UsageRecord' (Maybe Natural) (Maybe (NonEmpty UsageAllocation)) POSIX Text Text

newUsageRecord

Arguments

:: UTCTime
$sel:timestamp:UsageRecord'

-> Text
$sel:customerIdentifier:UsageRecord'

-> Text
$sel:dimension:UsageRecord'

-> UsageRecord

Create a value of UsageRecord with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:quantity:UsageRecord', usageRecord_quantity - The quantity of usage consumed by the customer for the given dimension and time. Defaults to 0 if not specified.

$sel:usageAllocations:UsageRecord', usageRecord_usageAllocations - The set of UsageAllocations to submit. The sum of all UsageAllocation quantities must equal the Quantity of the UsageRecord.

$sel:timestamp:UsageRecord', usageRecord_timestamp - Timestamp, in UTC, for which the usage is being reported.

Your application can meter usage for up to one hour in the past. Make sure the timestamp value is not before the start of the software usage.

$sel:customerIdentifier:UsageRecord', usageRecord_customerIdentifier - The CustomerIdentifier is obtained through the ResolveCustomer operation and represents an individual buyer in your application.

$sel:dimension:UsageRecord', usageRecord_dimension - During the process of registering a product on AWS Marketplace, up to eight dimensions are specified. These represent different units of value in your application.

UsageRecordResult

data UsageRecordResult

A UsageRecordResult indicates the status of a given UsageRecord processed by BatchMeterUsage.

See: newUsageRecordResult smart constructor.

Constructors

UsageRecordResult' (Maybe UsageRecordResultStatus) (Maybe UsageRecord) (Maybe Text)

newUsageRecordResult :: UsageRecordResult

Create a value of UsageRecordResult with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:status:UsageRecordResult', usageRecordResult_status - The UsageRecordResult Status indicates the status of an individual UsageRecord processed by BatchMeterUsage.

Success- The UsageRecord was accepted and honored by BatchMeterUsage.

CustomerNotSubscribed- The CustomerIdentifier specified is not subscribed to your product. The UsageRecord was not honored. Future UsageRecords for this customer will fail until the customer subscribes to your product.

DuplicateRecord- Indicates that the UsageRecord was invalid and not honored. A previously metered UsageRecord had the same customer, dimension, and time, but a different quantity.

$sel:usageRecord:UsageRecordResult', usageRecordResult_usageRecord - The UsageRecord that was part of the BatchMeterUsage request.

$sel:meteringRecordId:UsageRecordResult', usageRecordResult_meteringRecordId - The MeteringRecordId is a unique identifier for this metering event.