Amazonka.MarketplaceMetering.RegisterUsage

Description

Paid container software products sold through AWS Marketplace must integrate with the AWS Marketplace Metering Service and call the RegisterUsage operation for software entitlement and metering. Free and BYOL products for Amazon ECS or Amazon EKS aren't required to call RegisterUsage, but you may choose to do so if you would like to receive usage data in your seller reports. The sections below explain the behavior of RegisterUsage. RegisterUsage performs two primary functions: metering and entitlement.

Entitlement: RegisterUsage allows you to verify that the customer running your paid software is subscribed to your product on AWS Marketplace, enabling you to guard against unauthorized use. Your container image that integrates with RegisterUsage is only required to guard against unauthorized use at container startup, as such a CustomerNotSubscribedException/PlatformNotSupportedException will only be thrown on the initial call to RegisterUsage. Subsequent calls from the same Amazon ECS task instance (e.g. task-id) or Amazon EKS pod will not throw a CustomerNotSubscribedException, even if the customer unsubscribes while the Amazon ECS task or Amazon EKS pod is still running.

Metering: RegisterUsage meters software use per ECS task, per hour, or per pod for Amazon EKS with usage prorated to the second. A minimum of 1 minute of usage applies to tasks that are short lived. For example, if a customer has a 10 node Amazon ECS or Amazon EKS cluster and a service configured as a Daemon Set, then Amazon ECS or Amazon EKS will launch a task on all 10 cluster nodes and the customer will be charged: (10 * hourly_rate). Metering for software use is automatically handled by the AWS Marketplace Metering Control Plane -- your software is not required to perform any metering specific actions, other than call RegisterUsage once for metering of software use to commence. The AWS Marketplace Metering Control Plane will also continue to bill customers for running ECS tasks and Amazon EKS pods, regardless of the customers subscription state, removing the need for your software to perform entitlement checks at runtime.

Creating a Request

data RegisterUsage

See: newRegisterUsage smart constructor.

Constructors

RegisterUsage'

Fields

nonce :: Maybe Text

(Optional) To scope down the registration to a specific running software instance and guard against replay attacks.

productCode :: Text

Product code is used to uniquely identify a product in AWS Marketplace. The product code should be the same as the one used during the publishing of a new product.

publicKeyVersion :: Natural

Public Key Version provided by AWS Marketplace

newRegisterUsage

Arguments

:: Text
$sel:productCode:RegisterUsage'

-> Natural
$sel:publicKeyVersion:RegisterUsage'

-> RegisterUsage

Create a value of RegisterUsage with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:nonce:RegisterUsage', registerUsage_nonce - (Optional) To scope down the registration to a specific running software instance and guard against replay attacks.

$sel:productCode:RegisterUsage', registerUsage_productCode - Product code is used to uniquely identify a product in AWS Marketplace. The product code should be the same as the one used during the publishing of a new product.

$sel:publicKeyVersion:RegisterUsage', registerUsage_publicKeyVersion - Public Key Version provided by AWS Marketplace

Request Lenses

registerUsage_nonce :: Lens' RegisterUsage (Maybe Text)

(Optional) To scope down the registration to a specific running software instance and guard against replay attacks.

registerUsage_productCode :: Lens' RegisterUsage Text

Product code is used to uniquely identify a product in AWS Marketplace. The product code should be the same as the one used during the publishing of a new product.

registerUsage_publicKeyVersion :: Lens' RegisterUsage Natural

Public Key Version provided by AWS Marketplace

Destructuring the Response

data RegisterUsageResponse

See: newRegisterUsageResponse smart constructor.

Constructors

RegisterUsageResponse'

Fields

signature :: Maybe Text

JWT Token

publicKeyRotationTimestamp :: Maybe POSIX

(Optional) Only included when public key version has expired

httpStatus :: Int

The response's http status code.

newRegisterUsageResponse

Arguments

:: Int
$sel:httpStatus:RegisterUsageResponse'

-> RegisterUsageResponse

Create a value of RegisterUsageResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:signature:RegisterUsageResponse', registerUsageResponse_signature - JWT Token

$sel:publicKeyRotationTimestamp:RegisterUsageResponse', registerUsageResponse_publicKeyRotationTimestamp - (Optional) Only included when public key version has expired

$sel:httpStatus:RegisterUsageResponse', registerUsageResponse_httpStatus - The response's http status code.

Response Lenses

registerUsageResponse_signature :: Lens' RegisterUsageResponse (Maybe Text)

JWT Token

registerUsageResponse_publicKeyRotationTimestamp :: Lens' RegisterUsageResponse (Maybe UTCTime)

(Optional) Only included when public key version has expired

registerUsageResponse_httpStatus :: Lens' RegisterUsageResponse Int

The response's http status code.