Amazonka.MechanicalTurk.CreateHIT

Description

The CreateHIT operation creates a new Human Intelligence Task (HIT). The new HIT is made available for Workers to find and accept on the Amazon Mechanical Turk website.

This operation allows you to specify a new HIT by passing in values for the properties of the HIT, such as its title, reward amount and number of assignments. When you pass these values to CreateHIT, a new HIT is created for you, with a new HITTypeID. The HITTypeID can be used to create additional HITs in the future without needing to specify common parameters such as the title, description and reward amount each time.

An alternative way to create HITs is to first generate a HITTypeID using the CreateHITType operation and then call the CreateHITWithHITType operation. This is the recommended best practice for Requesters who are creating large numbers of HITs.

CreateHIT also supports several ways to provide question data: by providing a value for the Question parameter that fully specifies the contents of the HIT, or by providing a HitLayoutId and associated HitLayoutParameters.

If a HIT is created with 10 or more maximum assignments, there is an additional fee. For more information, see Amazon Mechanical Turk Pricing.

Creating a Request

data CreateHIT

See: newCreateHIT smart constructor.

Constructors

CreateHIT'

Fields

hITReviewPolicy :: Maybe ReviewPolicy

The HIT-level Review Policy applies to the HIT. You can specify for Mechanical Turk to take various actions based on the policy.

uniqueRequestToken :: Maybe Text

A unique identifier for this request which allows you to retry the call on error without creating duplicate HITs. This is useful in cases such as network timeouts where it is unclear whether or not the call succeeded on the server. If the HIT already exists in the system from a previous call using the same UniqueRequestToken, subsequent calls will return a AWS.MechanicalTurk.HitAlreadyExists error with a message containing the HITId.

Note: It is your responsibility to ensure uniqueness of the token. The unique token expires after 24 hours. Subsequent calls using the same UniqueRequestToken made after the 24 hour limit could create duplicate HITs.

autoApprovalDelayInSeconds :: Maybe Integer

The number of seconds after an assignment for the HIT has been submitted, after which the assignment is considered Approved automatically unless the Requester explicitly rejects it.

requesterAnnotation :: Maybe Text

An arbitrary data field. The RequesterAnnotation parameter lets your application attach arbitrary data to the HIT for tracking purposes. For example, this parameter could be an identifier internal to the Requester's application that corresponds with the HIT.

The RequesterAnnotation parameter for a HIT is only visible to the Requester who created the HIT. It is not shown to the Worker, or any other Requester.

The RequesterAnnotation parameter may be different for each HIT you submit. It does not affect how your HITs are grouped.

maxAssignments :: Maybe Int

The number of times the HIT can be accepted and completed before the HIT becomes unavailable.

keywords :: Maybe Text

One or more words or phrases that describe the HIT, separated by commas. These words are used in searches to find HITs.

hITLayoutId :: Maybe Text

The HITLayoutId allows you to use a pre-existing HIT design with placeholder values and create an additional HIT by providing those values as HITLayoutParameters.

Constraints: Either a Question parameter or a HITLayoutId parameter must be provided.

hITLayoutParameters :: Maybe [HITLayoutParameter]

If the HITLayoutId is provided, any placeholder values must be filled in with values using the HITLayoutParameter structure. For more information, see HITLayout.

qualificationRequirements :: Maybe [QualificationRequirement]

Conditions that a Worker's Qualifications must meet in order to accept the HIT. A HIT can have between zero and ten Qualification requirements. All requirements must be met in order for a Worker to accept the HIT. Additionally, other actions can be restricted using the ActionsGuarded field on each QualificationRequirement structure.

question :: Maybe Text

The data the person completing the HIT uses to produce the results.

Constraints: Must be a QuestionForm data structure, an ExternalQuestion data structure, or an HTMLQuestion data structure. The XML question data must not be larger than 64 kilobytes (65,535 bytes) in size, including whitespace.

Either a Question parameter or a HITLayoutId parameter must be provided.

assignmentReviewPolicy :: Maybe ReviewPolicy

The Assignment-level Review Policy applies to the assignments under the HIT. You can specify for Mechanical Turk to take various actions based on the policy.

lifetimeInSeconds :: Integer

An amount of time, in seconds, after which the HIT is no longer available for users to accept. After the lifetime of the HIT elapses, the HIT no longer appears in HIT searches, even if not all of the assignments for the HIT have been accepted.

assignmentDurationInSeconds :: Integer

The amount of time, in seconds, that a Worker has to complete the HIT after accepting it. If a Worker does not complete the assignment within the specified duration, the assignment is considered abandoned. If the HIT is still active (that is, its lifetime has not elapsed), the assignment becomes available for other users to find and accept.

reward :: Text

The amount of money the Requester will pay a Worker for successfully completing the HIT.

title :: Text

The title of the HIT. A title should be short and descriptive about the kind of task the HIT contains. On the Amazon Mechanical Turk web site, the HIT title appears in search results, and everywhere the HIT is mentioned.

description :: Text

A general description of the HIT. A description includes detailed information about the kind of task the HIT contains. On the Amazon Mechanical Turk web site, the HIT description appears in the expanded view of search results, and in the HIT and assignment screens. A good description gives the user enough information to evaluate the HIT before accepting it.

newCreateHIT

Arguments

:: Integer
$sel:lifetimeInSeconds:CreateHIT'

-> Integer
$sel:assignmentDurationInSeconds:CreateHIT'

-> Text
$sel:reward:CreateHIT'

-> Text
$sel:title:CreateHIT'

-> Text
$sel:description:CreateHIT'

-> CreateHIT

Create a value of CreateHIT with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:hITReviewPolicy:CreateHIT', createHIT_hITReviewPolicy - The HIT-level Review Policy applies to the HIT. You can specify for Mechanical Turk to take various actions based on the policy.

$sel:uniqueRequestToken:CreateHIT', createHIT_uniqueRequestToken - A unique identifier for this request which allows you to retry the call on error without creating duplicate HITs. This is useful in cases such as network timeouts where it is unclear whether or not the call succeeded on the server. If the HIT already exists in the system from a previous call using the same UniqueRequestToken, subsequent calls will return a AWS.MechanicalTurk.HitAlreadyExists error with a message containing the HITId.

Note: It is your responsibility to ensure uniqueness of the token. The unique token expires after 24 hours. Subsequent calls using the same UniqueRequestToken made after the 24 hour limit could create duplicate HITs.

$sel:autoApprovalDelayInSeconds:CreateHIT', createHIT_autoApprovalDelayInSeconds - The number of seconds after an assignment for the HIT has been submitted, after which the assignment is considered Approved automatically unless the Requester explicitly rejects it.

$sel:requesterAnnotation:CreateHIT', createHIT_requesterAnnotation - An arbitrary data field. The RequesterAnnotation parameter lets your application attach arbitrary data to the HIT for tracking purposes. For example, this parameter could be an identifier internal to the Requester's application that corresponds with the HIT.

The RequesterAnnotation parameter for a HIT is only visible to the Requester who created the HIT. It is not shown to the Worker, or any other Requester.

The RequesterAnnotation parameter may be different for each HIT you submit. It does not affect how your HITs are grouped.

$sel:maxAssignments:CreateHIT', createHIT_maxAssignments - The number of times the HIT can be accepted and completed before the HIT becomes unavailable.

$sel:keywords:CreateHIT', createHIT_keywords - One or more words or phrases that describe the HIT, separated by commas. These words are used in searches to find HITs.

$sel:hITLayoutId:CreateHIT', createHIT_hITLayoutId - The HITLayoutId allows you to use a pre-existing HIT design with placeholder values and create an additional HIT by providing those values as HITLayoutParameters.

Constraints: Either a Question parameter or a HITLayoutId parameter must be provided.

$sel:hITLayoutParameters:CreateHIT', createHIT_hITLayoutParameters - If the HITLayoutId is provided, any placeholder values must be filled in with values using the HITLayoutParameter structure. For more information, see HITLayout.

$sel:qualificationRequirements:CreateHIT', createHIT_qualificationRequirements - Conditions that a Worker's Qualifications must meet in order to accept the HIT. A HIT can have between zero and ten Qualification requirements. All requirements must be met in order for a Worker to accept the HIT. Additionally, other actions can be restricted using the ActionsGuarded field on each QualificationRequirement structure.

$sel:question:CreateHIT', createHIT_question - The data the person completing the HIT uses to produce the results.

Constraints: Must be a QuestionForm data structure, an ExternalQuestion data structure, or an HTMLQuestion data structure. The XML question data must not be larger than 64 kilobytes (65,535 bytes) in size, including whitespace.

Either a Question parameter or a HITLayoutId parameter must be provided.

$sel:assignmentReviewPolicy:CreateHIT', createHIT_assignmentReviewPolicy - The Assignment-level Review Policy applies to the assignments under the HIT. You can specify for Mechanical Turk to take various actions based on the policy.

$sel:lifetimeInSeconds:CreateHIT', createHIT_lifetimeInSeconds - An amount of time, in seconds, after which the HIT is no longer available for users to accept. After the lifetime of the HIT elapses, the HIT no longer appears in HIT searches, even if not all of the assignments for the HIT have been accepted.

$sel:assignmentDurationInSeconds:CreateHIT', createHIT_assignmentDurationInSeconds - The amount of time, in seconds, that a Worker has to complete the HIT after accepting it. If a Worker does not complete the assignment within the specified duration, the assignment is considered abandoned. If the HIT is still active (that is, its lifetime has not elapsed), the assignment becomes available for other users to find and accept.

$sel:reward:CreateHIT', createHIT_reward - The amount of money the Requester will pay a Worker for successfully completing the HIT.

$sel:title:CreateHIT', createHIT_title - The title of the HIT. A title should be short and descriptive about the kind of task the HIT contains. On the Amazon Mechanical Turk web site, the HIT title appears in search results, and everywhere the HIT is mentioned.

$sel:description:CreateHIT', createHIT_description - A general description of the HIT. A description includes detailed information about the kind of task the HIT contains. On the Amazon Mechanical Turk web site, the HIT description appears in the expanded view of search results, and in the HIT and assignment screens. A good description gives the user enough information to evaluate the HIT before accepting it.

Request Lenses

createHIT_hITReviewPolicy :: Lens' CreateHIT (Maybe ReviewPolicy)

The HIT-level Review Policy applies to the HIT. You can specify for Mechanical Turk to take various actions based on the policy.

createHIT_uniqueRequestToken :: Lens' CreateHIT (Maybe Text)

A unique identifier for this request which allows you to retry the call on error without creating duplicate HITs. This is useful in cases such as network timeouts where it is unclear whether or not the call succeeded on the server. If the HIT already exists in the system from a previous call using the same UniqueRequestToken, subsequent calls will return a AWS.MechanicalTurk.HitAlreadyExists error with a message containing the HITId.

Note: It is your responsibility to ensure uniqueness of the token. The unique token expires after 24 hours. Subsequent calls using the same UniqueRequestToken made after the 24 hour limit could create duplicate HITs.

createHIT_autoApprovalDelayInSeconds :: Lens' CreateHIT (Maybe Integer)

The number of seconds after an assignment for the HIT has been submitted, after which the assignment is considered Approved automatically unless the Requester explicitly rejects it.

createHIT_requesterAnnotation :: Lens' CreateHIT (Maybe Text)

An arbitrary data field. The RequesterAnnotation parameter lets your application attach arbitrary data to the HIT for tracking purposes. For example, this parameter could be an identifier internal to the Requester's application that corresponds with the HIT.

The RequesterAnnotation parameter for a HIT is only visible to the Requester who created the HIT. It is not shown to the Worker, or any other Requester.

The RequesterAnnotation parameter may be different for each HIT you submit. It does not affect how your HITs are grouped.

createHIT_maxAssignments :: Lens' CreateHIT (Maybe Int)

The number of times the HIT can be accepted and completed before the HIT becomes unavailable.

createHIT_keywords :: Lens' CreateHIT (Maybe Text)

One or more words or phrases that describe the HIT, separated by commas. These words are used in searches to find HITs.

createHIT_hITLayoutId :: Lens' CreateHIT (Maybe Text)

The HITLayoutId allows you to use a pre-existing HIT design with placeholder values and create an additional HIT by providing those values as HITLayoutParameters.

Constraints: Either a Question parameter or a HITLayoutId parameter must be provided.

createHIT_hITLayoutParameters :: Lens' CreateHIT (Maybe [HITLayoutParameter])

If the HITLayoutId is provided, any placeholder values must be filled in with values using the HITLayoutParameter structure. For more information, see HITLayout.

createHIT_qualificationRequirements :: Lens' CreateHIT (Maybe [QualificationRequirement])

Conditions that a Worker's Qualifications must meet in order to accept the HIT. A HIT can have between zero and ten Qualification requirements. All requirements must be met in order for a Worker to accept the HIT. Additionally, other actions can be restricted using the ActionsGuarded field on each QualificationRequirement structure.

createHIT_question :: Lens' CreateHIT (Maybe Text)

The data the person completing the HIT uses to produce the results.

Constraints: Must be a QuestionForm data structure, an ExternalQuestion data structure, or an HTMLQuestion data structure. The XML question data must not be larger than 64 kilobytes (65,535 bytes) in size, including whitespace.

Either a Question parameter or a HITLayoutId parameter must be provided.

createHIT_assignmentReviewPolicy :: Lens' CreateHIT (Maybe ReviewPolicy)

The Assignment-level Review Policy applies to the assignments under the HIT. You can specify for Mechanical Turk to take various actions based on the policy.

createHIT_lifetimeInSeconds :: Lens' CreateHIT Integer

An amount of time, in seconds, after which the HIT is no longer available for users to accept. After the lifetime of the HIT elapses, the HIT no longer appears in HIT searches, even if not all of the assignments for the HIT have been accepted.

createHIT_assignmentDurationInSeconds :: Lens' CreateHIT Integer

The amount of time, in seconds, that a Worker has to complete the HIT after accepting it. If a Worker does not complete the assignment within the specified duration, the assignment is considered abandoned. If the HIT is still active (that is, its lifetime has not elapsed), the assignment becomes available for other users to find and accept.

createHIT_reward :: Lens' CreateHIT Text

The amount of money the Requester will pay a Worker for successfully completing the HIT.

createHIT_title :: Lens' CreateHIT Text

The title of the HIT. A title should be short and descriptive about the kind of task the HIT contains. On the Amazon Mechanical Turk web site, the HIT title appears in search results, and everywhere the HIT is mentioned.

createHIT_description :: Lens' CreateHIT Text

A general description of the HIT. A description includes detailed information about the kind of task the HIT contains. On the Amazon Mechanical Turk web site, the HIT description appears in the expanded view of search results, and in the HIT and assignment screens. A good description gives the user enough information to evaluate the HIT before accepting it.

Destructuring the Response

data CreateHITResponse

See: newCreateHITResponse smart constructor.

Constructors

CreateHITResponse'

Fields

hit :: Maybe HIT

Contains the newly created HIT data. For a description of the HIT data structure as it appears in responses, see the HIT Data Structure documentation.

httpStatus :: Int

The response's http status code.

newCreateHITResponse

Arguments

:: Int
$sel:httpStatus:CreateHITResponse'

-> CreateHITResponse

Create a value of CreateHITResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:hit:CreateHITResponse', createHITResponse_hit - Contains the newly created HIT data. For a description of the HIT data structure as it appears in responses, see the HIT Data Structure documentation.

$sel:httpStatus:CreateHITResponse', createHITResponse_httpStatus - The response's http status code.

Response Lenses

createHITResponse_hit :: Lens' CreateHITResponse (Maybe HIT)

Contains the newly created HIT data. For a description of the HIT data structure as it appears in responses, see the HIT Data Structure documentation.

createHITResponse_httpStatus :: Lens' CreateHITResponse Int

The response's http status code.