Amazonka.MechanicalTurk.CreateQualificationType

Description

The CreateQualificationType operation creates a new Qualification type, which is represented by a QualificationType data structure.

Creating a Request

data CreateQualificationType

See: newCreateQualificationType smart constructor.

Constructors

CreateQualificationType'

Fields

testDurationInSeconds :: Maybe Integer

The number of seconds the Worker has to complete the Qualification test, starting from the time the Worker requests the Qualification.

answerKey :: Maybe Text

The answers to the Qualification test specified in the Test parameter, in the form of an AnswerKey data structure.

Constraints: Must not be longer than 65535 bytes.

Constraints: None. If not specified, you must process Qualification requests manually.

test :: Maybe Text

The questions for the Qualification test a Worker must answer correctly to obtain a Qualification of this type. If this parameter is specified, TestDurationInSeconds must also be specified.

Constraints: Must not be longer than 65535 bytes. Must be a QuestionForm data structure. This parameter cannot be specified if AutoGranted is true.

Constraints: None. If not specified, the Worker may request the Qualification without answering any questions.

keywords :: Maybe Text

One or more words or phrases that describe the Qualification type, separated by commas. The keywords of a type make the type easier to find during a search.

autoGranted :: Maybe Bool

Specifies whether requests for the Qualification type are granted immediately, without prompting the Worker with a Qualification test.

Constraints: If the Test parameter is specified, this parameter cannot be true.

autoGrantedValue :: Maybe Int

The Qualification value to use for automatically granted Qualifications. This parameter is used only if the AutoGranted parameter is true.

retryDelayInSeconds :: Maybe Integer

The number of seconds that a Worker must wait after requesting a Qualification of the Qualification type before the worker can retry the Qualification request.

Constraints: None. If not specified, retries are disabled and Workers can request a Qualification of this type only once, even if the Worker has not been granted the Qualification. It is not possible to disable retries for a Qualification type after it has been created with retries enabled. If you want to disable retries, you must delete existing retry-enabled Qualification type and then create a new Qualification type with retries disabled.

name :: Text

The name you give to the Qualification type. The type name is used to represent the Qualification to Workers, and to find the type using a Qualification type search. It must be unique across all of your Qualification types.

description :: Text

A long description for the Qualification type. On the Amazon Mechanical Turk website, the long description is displayed when a Worker examines a Qualification type.

qualificationTypeStatus :: QualificationTypeStatus

The initial status of the Qualification type.

Constraints: Valid values are: Active | Inactive

newCreateQualificationType

Arguments

:: Text
$sel:name:CreateQualificationType'

-> Text
$sel:description:CreateQualificationType'

-> QualificationTypeStatus
$sel:qualificationTypeStatus:CreateQualificationType'

-> CreateQualificationType

Create a value of CreateQualificationType with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:testDurationInSeconds:CreateQualificationType', createQualificationType_testDurationInSeconds - The number of seconds the Worker has to complete the Qualification test, starting from the time the Worker requests the Qualification.

$sel:answerKey:CreateQualificationType', createQualificationType_answerKey - The answers to the Qualification test specified in the Test parameter, in the form of an AnswerKey data structure.

Constraints: Must not be longer than 65535 bytes.

Constraints: None. If not specified, you must process Qualification requests manually.

$sel:test:CreateQualificationType', createQualificationType_test - The questions for the Qualification test a Worker must answer correctly to obtain a Qualification of this type. If this parameter is specified, TestDurationInSeconds must also be specified.

Constraints: Must not be longer than 65535 bytes. Must be a QuestionForm data structure. This parameter cannot be specified if AutoGranted is true.

Constraints: None. If not specified, the Worker may request the Qualification without answering any questions.

$sel:keywords:CreateQualificationType', createQualificationType_keywords - One or more words or phrases that describe the Qualification type, separated by commas. The keywords of a type make the type easier to find during a search.

$sel:autoGranted:CreateQualificationType', createQualificationType_autoGranted - Specifies whether requests for the Qualification type are granted immediately, without prompting the Worker with a Qualification test.

Constraints: If the Test parameter is specified, this parameter cannot be true.

$sel:autoGrantedValue:CreateQualificationType', createQualificationType_autoGrantedValue - The Qualification value to use for automatically granted Qualifications. This parameter is used only if the AutoGranted parameter is true.

$sel:retryDelayInSeconds:CreateQualificationType', createQualificationType_retryDelayInSeconds - The number of seconds that a Worker must wait after requesting a Qualification of the Qualification type before the worker can retry the Qualification request.

Constraints: None. If not specified, retries are disabled and Workers can request a Qualification of this type only once, even if the Worker has not been granted the Qualification. It is not possible to disable retries for a Qualification type after it has been created with retries enabled. If you want to disable retries, you must delete existing retry-enabled Qualification type and then create a new Qualification type with retries disabled.

$sel:name:CreateQualificationType', createQualificationType_name - The name you give to the Qualification type. The type name is used to represent the Qualification to Workers, and to find the type using a Qualification type search. It must be unique across all of your Qualification types.

$sel:description:CreateQualificationType', createQualificationType_description - A long description for the Qualification type. On the Amazon Mechanical Turk website, the long description is displayed when a Worker examines a Qualification type.

$sel:qualificationTypeStatus:CreateQualificationType', createQualificationType_qualificationTypeStatus - The initial status of the Qualification type.

Constraints: Valid values are: Active | Inactive

Request Lenses

createQualificationType_testDurationInSeconds :: Lens' CreateQualificationType (Maybe Integer)

The number of seconds the Worker has to complete the Qualification test, starting from the time the Worker requests the Qualification.

createQualificationType_answerKey :: Lens' CreateQualificationType (Maybe Text)

The answers to the Qualification test specified in the Test parameter, in the form of an AnswerKey data structure.

Constraints: Must not be longer than 65535 bytes.

Constraints: None. If not specified, you must process Qualification requests manually.

createQualificationType_test :: Lens' CreateQualificationType (Maybe Text)

The questions for the Qualification test a Worker must answer correctly to obtain a Qualification of this type. If this parameter is specified, TestDurationInSeconds must also be specified.

Constraints: Must not be longer than 65535 bytes. Must be a QuestionForm data structure. This parameter cannot be specified if AutoGranted is true.

Constraints: None. If not specified, the Worker may request the Qualification without answering any questions.

createQualificationType_keywords :: Lens' CreateQualificationType (Maybe Text)

One or more words or phrases that describe the Qualification type, separated by commas. The keywords of a type make the type easier to find during a search.

createQualificationType_autoGranted :: Lens' CreateQualificationType (Maybe Bool)

Specifies whether requests for the Qualification type are granted immediately, without prompting the Worker with a Qualification test.

Constraints: If the Test parameter is specified, this parameter cannot be true.

createQualificationType_autoGrantedValue :: Lens' CreateQualificationType (Maybe Int)

The Qualification value to use for automatically granted Qualifications. This parameter is used only if the AutoGranted parameter is true.

createQualificationType_retryDelayInSeconds :: Lens' CreateQualificationType (Maybe Integer)

The number of seconds that a Worker must wait after requesting a Qualification of the Qualification type before the worker can retry the Qualification request.

Constraints: None. If not specified, retries are disabled and Workers can request a Qualification of this type only once, even if the Worker has not been granted the Qualification. It is not possible to disable retries for a Qualification type after it has been created with retries enabled. If you want to disable retries, you must delete existing retry-enabled Qualification type and then create a new Qualification type with retries disabled.

createQualificationType_name :: Lens' CreateQualificationType Text

The name you give to the Qualification type. The type name is used to represent the Qualification to Workers, and to find the type using a Qualification type search. It must be unique across all of your Qualification types.

createQualificationType_description :: Lens' CreateQualificationType Text

A long description for the Qualification type. On the Amazon Mechanical Turk website, the long description is displayed when a Worker examines a Qualification type.

createQualificationType_qualificationTypeStatus :: Lens' CreateQualificationType QualificationTypeStatus

The initial status of the Qualification type.

Constraints: Valid values are: Active | Inactive

Destructuring the Response

data CreateQualificationTypeResponse

See: newCreateQualificationTypeResponse smart constructor.

Constructors

CreateQualificationTypeResponse'

Fields

qualificationType :: Maybe QualificationType

The created Qualification type, returned as a QualificationType data structure.

httpStatus :: Int

The response's http status code.

newCreateQualificationTypeResponse

Arguments

:: Int
$sel:httpStatus:CreateQualificationTypeResponse'

-> CreateQualificationTypeResponse

Create a value of CreateQualificationTypeResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:qualificationType:CreateQualificationTypeResponse', createQualificationTypeResponse_qualificationType - The created Qualification type, returned as a QualificationType data structure.

$sel:httpStatus:CreateQualificationTypeResponse', createQualificationTypeResponse_httpStatus - The response's http status code.

Response Lenses

createQualificationTypeResponse_qualificationType :: Lens' CreateQualificationTypeResponse (Maybe QualificationType)

The created Qualification type, returned as a QualificationType data structure.

createQualificationTypeResponse_httpStatus :: Lens' CreateQualificationTypeResponse Int

The response's http status code.