Amazonka.MechanicalTurk.ListQualificationTypes

Description

The ListQualificationTypes operation returns a list of Qualification types, filtered by an optional search term.

This operation returns paginated results.

Creating a Request

data ListQualificationTypes

See: newListQualificationTypes smart constructor.

Constructors

ListQualificationTypes'

Fields

mustBeOwnedByCaller :: Maybe Bool

Specifies that only Qualification types that the Requester created are returned. If false, the operation returns all Qualification types.

nextToken :: Maybe Text

query :: Maybe Text

A text query against all of the searchable attributes of Qualification types.

maxResults :: Maybe Natural

The maximum number of results to return in a single call.

mustBeRequestable :: Bool

Specifies that only Qualification types that a user can request through the Amazon Mechanical Turk web site, such as by taking a Qualification test, are returned as results of the search. Some Qualification types, such as those assigned automatically by the system, cannot be requested directly by users. If false, all Qualification types, including those managed by the system, are considered. Valid values are True | False.

newListQualificationTypes

Arguments

:: Bool
$sel:mustBeRequestable:ListQualificationTypes'

-> ListQualificationTypes

Create a value of ListQualificationTypes with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:mustBeOwnedByCaller:ListQualificationTypes', listQualificationTypes_mustBeOwnedByCaller - Specifies that only Qualification types that the Requester created are returned. If false, the operation returns all Qualification types.

$sel:nextToken:ListQualificationTypes', listQualificationTypes_nextToken - Undocumented member.

$sel:query:ListQualificationTypes', listQualificationTypes_query - A text query against all of the searchable attributes of Qualification types.

$sel:maxResults:ListQualificationTypes', listQualificationTypes_maxResults - The maximum number of results to return in a single call.

$sel:mustBeRequestable:ListQualificationTypes', listQualificationTypes_mustBeRequestable - Specifies that only Qualification types that a user can request through the Amazon Mechanical Turk web site, such as by taking a Qualification test, are returned as results of the search. Some Qualification types, such as those assigned automatically by the system, cannot be requested directly by users. If false, all Qualification types, including those managed by the system, are considered. Valid values are True | False.

Request Lenses

listQualificationTypes_mustBeOwnedByCaller :: Lens' ListQualificationTypes (Maybe Bool)

Specifies that only Qualification types that the Requester created are returned. If false, the operation returns all Qualification types.

listQualificationTypes_nextToken :: Lens' ListQualificationTypes (Maybe Text)

Undocumented member.

listQualificationTypes_query :: Lens' ListQualificationTypes (Maybe Text)

A text query against all of the searchable attributes of Qualification types.

listQualificationTypes_maxResults :: Lens' ListQualificationTypes (Maybe Natural)

The maximum number of results to return in a single call.

listQualificationTypes_mustBeRequestable :: Lens' ListQualificationTypes Bool

Specifies that only Qualification types that a user can request through the Amazon Mechanical Turk web site, such as by taking a Qualification test, are returned as results of the search. Some Qualification types, such as those assigned automatically by the system, cannot be requested directly by users. If false, all Qualification types, including those managed by the system, are considered. Valid values are True | False.

Destructuring the Response

data ListQualificationTypesResponse

See: newListQualificationTypesResponse smart constructor.

Constructors

ListQualificationTypesResponse'

Fields

qualificationTypes :: Maybe [QualificationType]

The list of QualificationType elements returned by the query.

nextToken :: Maybe Text

numResults :: Maybe Int

The number of Qualification types on this page in the filtered results list, equivalent to the number of types this operation returns.

httpStatus :: Int

The response's http status code.

newListQualificationTypesResponse

Arguments

:: Int
$sel:httpStatus:ListQualificationTypesResponse'

-> ListQualificationTypesResponse

Create a value of ListQualificationTypesResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:qualificationTypes:ListQualificationTypesResponse', listQualificationTypesResponse_qualificationTypes - The list of QualificationType elements returned by the query.

$sel:nextToken:ListQualificationTypes', listQualificationTypesResponse_nextToken - Undocumented member.

$sel:numResults:ListQualificationTypesResponse', listQualificationTypesResponse_numResults - The number of Qualification types on this page in the filtered results list, equivalent to the number of types this operation returns.

$sel:httpStatus:ListQualificationTypesResponse', listQualificationTypesResponse_httpStatus - The response's http status code.

Response Lenses

listQualificationTypesResponse_qualificationTypes :: Lens' ListQualificationTypesResponse (Maybe [QualificationType])

The list of QualificationType elements returned by the query.

listQualificationTypesResponse_nextToken :: Lens' ListQualificationTypesResponse (Maybe Text)

Undocumented member.

listQualificationTypesResponse_numResults :: Lens' ListQualificationTypesResponse (Maybe Int)

The number of Qualification types on this page in the filtered results list, equivalent to the number of types this operation returns.

listQualificationTypesResponse_httpStatus :: Lens' ListQualificationTypesResponse Int

The response's http status code.