Amazonka.MechanicalTurk.NotifyWorkers

Description

The NotifyWorkers operation sends an email to one or more Workers that you specify with the Worker ID. You can specify up to 100 Worker IDs to send the same message with a single call to the NotifyWorkers operation. The NotifyWorkers operation will send a notification email to a Worker only if you have previously approved or rejected work from the Worker.

Creating a Request

data NotifyWorkers

See: newNotifyWorkers smart constructor.

Constructors

NotifyWorkers'

Fields

subject :: Text

The subject line of the email message to send. Can include up to 200 characters.

messageText :: Text

The text of the email message to send. Can include up to 4,096 characters

workerIds :: [Text]

A list of Worker IDs you wish to notify. You can notify upto 100 Workers at a time.

newNotifyWorkers

Arguments

:: Text
$sel:subject:NotifyWorkers'

-> Text
$sel:messageText:NotifyWorkers'

-> NotifyWorkers

Create a value of NotifyWorkers with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:subject:NotifyWorkers', notifyWorkers_subject - The subject line of the email message to send. Can include up to 200 characters.

$sel:messageText:NotifyWorkers', notifyWorkers_messageText - The text of the email message to send. Can include up to 4,096 characters

$sel:workerIds:NotifyWorkers', notifyWorkers_workerIds - A list of Worker IDs you wish to notify. You can notify upto 100 Workers at a time.

Request Lenses

notifyWorkers_subject :: Lens' NotifyWorkers Text

The subject line of the email message to send. Can include up to 200 characters.

notifyWorkers_messageText :: Lens' NotifyWorkers Text

The text of the email message to send. Can include up to 4,096 characters

notifyWorkers_workerIds :: Lens' NotifyWorkers [Text]

A list of Worker IDs you wish to notify. You can notify upto 100 Workers at a time.

Destructuring the Response

data NotifyWorkersResponse

See: newNotifyWorkersResponse smart constructor.

Constructors

NotifyWorkersResponse'

Fields

notifyWorkersFailureStatuses :: Maybe [NotifyWorkersFailureStatus]

When MTurk sends notifications to the list of Workers, it returns back any failures it encounters in this list of NotifyWorkersFailureStatus objects.

httpStatus :: Int

The response's http status code.

newNotifyWorkersResponse

Arguments

:: Int
$sel:httpStatus:NotifyWorkersResponse'

-> NotifyWorkersResponse

Create a value of NotifyWorkersResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:notifyWorkersFailureStatuses:NotifyWorkersResponse', notifyWorkersResponse_notifyWorkersFailureStatuses - When MTurk sends notifications to the list of Workers, it returns back any failures it encounters in this list of NotifyWorkersFailureStatus objects.

$sel:httpStatus:NotifyWorkersResponse', notifyWorkersResponse_httpStatus - The response's http status code.

Response Lenses

notifyWorkersResponse_notifyWorkersFailureStatuses :: Lens' NotifyWorkersResponse (Maybe [NotifyWorkersFailureStatus])

When MTurk sends notifications to the list of Workers, it returns back any failures it encounters in this list of NotifyWorkersFailureStatus objects.

notifyWorkersResponse_httpStatus :: Lens' NotifyWorkersResponse Int

The response's http status code.