Amazonka.MechanicalTurk.Types.QualificationRequest

Description

Documentation

data QualificationRequest

The QualificationRequest data structure represents a request a Worker has made for a Qualification.

See: newQualificationRequest smart constructor.

Constructors

QualificationRequest'

Fields

qualificationRequestId :: Maybe Text

The ID of the Qualification request, a unique identifier generated when the request was submitted.

test :: Maybe Text

The contents of the Qualification test that was presented to the Worker, if the type has a test and the Worker has submitted answers. This value is identical to the QuestionForm associated with the Qualification type at the time the Worker requests the Qualification.

qualificationTypeId :: Maybe Text

The ID of the Qualification type the Worker is requesting, as returned by the CreateQualificationType operation.

answer :: Maybe Text

The Worker's answers for the Qualification type's test contained in a QuestionFormAnswers document, if the type has a test and the Worker has submitted answers. If the Worker does not provide any answers, Answer may be empty.

workerId :: Maybe Text

The ID of the Worker requesting the Qualification.

submitTime :: Maybe POSIX

The date and time the Qualification request had a status of Submitted. This is either the time the Worker submitted answers for a Qualification test, or the time the Worker requested the Qualification if the Qualification type does not have a test.

newQualificationRequest :: QualificationRequest

Create a value of QualificationRequest with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:qualificationRequestId:QualificationRequest', qualificationRequest_qualificationRequestId - The ID of the Qualification request, a unique identifier generated when the request was submitted.

$sel:test:QualificationRequest', qualificationRequest_test - The contents of the Qualification test that was presented to the Worker, if the type has a test and the Worker has submitted answers. This value is identical to the QuestionForm associated with the Qualification type at the time the Worker requests the Qualification.

$sel:qualificationTypeId:QualificationRequest', qualificationRequest_qualificationTypeId - The ID of the Qualification type the Worker is requesting, as returned by the CreateQualificationType operation.

$sel:answer:QualificationRequest', qualificationRequest_answer - The Worker's answers for the Qualification type's test contained in a QuestionFormAnswers document, if the type has a test and the Worker has submitted answers. If the Worker does not provide any answers, Answer may be empty.

$sel:workerId:QualificationRequest', qualificationRequest_workerId - The ID of the Worker requesting the Qualification.

$sel:submitTime:QualificationRequest', qualificationRequest_submitTime - The date and time the Qualification request had a status of Submitted. This is either the time the Worker submitted answers for a Qualification test, or the time the Worker requested the Qualification if the Qualification type does not have a test.

qualificationRequest_qualificationRequestId :: Lens' QualificationRequest (Maybe Text)

The ID of the Qualification request, a unique identifier generated when the request was submitted.

qualificationRequest_test :: Lens' QualificationRequest (Maybe Text)

The contents of the Qualification test that was presented to the Worker, if the type has a test and the Worker has submitted answers. This value is identical to the QuestionForm associated with the Qualification type at the time the Worker requests the Qualification.

qualificationRequest_qualificationTypeId :: Lens' QualificationRequest (Maybe Text)

The ID of the Qualification type the Worker is requesting, as returned by the CreateQualificationType operation.

qualificationRequest_answer :: Lens' QualificationRequest (Maybe Text)

The Worker's answers for the Qualification type's test contained in a QuestionFormAnswers document, if the type has a test and the Worker has submitted answers. If the Worker does not provide any answers, Answer may be empty.

qualificationRequest_workerId :: Lens' QualificationRequest (Maybe Text)

The ID of the Worker requesting the Qualification.

qualificationRequest_submitTime :: Lens' QualificationRequest (Maybe UTCTime)

The date and time the Qualification request had a status of Submitted. This is either the time the Worker submitted answers for a Qualification test, or the time the Worker requested the Qualification if the Qualification type does not have a test.