Amazonka.MechanicalTurk.UpdateHITTypeOfHIT

Description

The UpdateHITTypeOfHIT operation allows you to change the HITType properties of a HIT. This operation disassociates the HIT from its old HITType properties and associates it with the new HITType properties. The HIT takes on the properties of the new HITType in place of the old ones.

Creating a Request

data UpdateHITTypeOfHIT

See: newUpdateHITTypeOfHIT smart constructor.

Constructors

UpdateHITTypeOfHIT'

Fields

hITId :: Text

The HIT to update.

hITTypeId :: Text

The ID of the new HIT type.

newUpdateHITTypeOfHIT

Arguments

:: Text
$sel:hITId:UpdateHITTypeOfHIT'

-> Text
$sel:hITTypeId:UpdateHITTypeOfHIT'

-> UpdateHITTypeOfHIT

Create a value of UpdateHITTypeOfHIT with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:hITId:UpdateHITTypeOfHIT', updateHITTypeOfHIT_hITId - The HIT to update.

$sel:hITTypeId:UpdateHITTypeOfHIT', updateHITTypeOfHIT_hITTypeId - The ID of the new HIT type.

Request Lenses

updateHITTypeOfHIT_hITId :: Lens' UpdateHITTypeOfHIT Text

The HIT to update.

updateHITTypeOfHIT_hITTypeId :: Lens' UpdateHITTypeOfHIT Text

The ID of the new HIT type.

Destructuring the Response

data UpdateHITTypeOfHITResponse

See: newUpdateHITTypeOfHITResponse smart constructor.

Constructors

UpdateHITTypeOfHITResponse'

Fields

httpStatus :: Int

The response's http status code.

newUpdateHITTypeOfHITResponse

Arguments

:: Int
$sel:httpStatus:UpdateHITTypeOfHITResponse'

-> UpdateHITTypeOfHITResponse

Create a value of UpdateHITTypeOfHITResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:httpStatus:UpdateHITTypeOfHITResponse', updateHITTypeOfHITResponse_httpStatus - The response's http status code.

Response Lenses

updateHITTypeOfHITResponse_httpStatus :: Lens' UpdateHITTypeOfHITResponse Int

The response's http status code.