Amazonka.MediaConnect.RemoveFlowOutput

Description

Removes an output from an existing flow. This request can be made only on an output that does not have an entitlement associated with it. If the output has an entitlement, you must revoke the entitlement instead. When an entitlement is revoked from a flow, the service automatically removes the associated output.

Creating a Request

data RemoveFlowOutput

See: newRemoveFlowOutput smart constructor.

Constructors

RemoveFlowOutput'

Fields

flowArn :: Text

The flow that you want to remove an output from.

outputArn :: Text

The ARN of the output that you want to remove.

newRemoveFlowOutput

Arguments

:: Text
$sel:flowArn:RemoveFlowOutput'

-> Text
$sel:outputArn:RemoveFlowOutput'

-> RemoveFlowOutput

Create a value of RemoveFlowOutput with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:flowArn:RemoveFlowOutput', removeFlowOutput_flowArn - The flow that you want to remove an output from.

$sel:outputArn:RemoveFlowOutput', removeFlowOutput_outputArn - The ARN of the output that you want to remove.

Request Lenses

removeFlowOutput_flowArn :: Lens' RemoveFlowOutput Text

The flow that you want to remove an output from.

removeFlowOutput_outputArn :: Lens' RemoveFlowOutput Text

The ARN of the output that you want to remove.

Destructuring the Response

data RemoveFlowOutputResponse

See: newRemoveFlowOutputResponse smart constructor.

Constructors

RemoveFlowOutputResponse'

Fields

flowArn :: Maybe Text

The ARN of the flow that is associated with the output you removed.

outputArn :: Maybe Text

The ARN of the output that was removed.

httpStatus :: Int

The response's http status code.

newRemoveFlowOutputResponse

Arguments

:: Int
$sel:httpStatus:RemoveFlowOutputResponse'

-> RemoveFlowOutputResponse

Create a value of RemoveFlowOutputResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:flowArn:RemoveFlowOutput', removeFlowOutputResponse_flowArn - The ARN of the flow that is associated with the output you removed.

$sel:outputArn:RemoveFlowOutput', removeFlowOutputResponse_outputArn - The ARN of the output that was removed.

$sel:httpStatus:RemoveFlowOutputResponse', removeFlowOutputResponse_httpStatus - The response's http status code.

Response Lenses

removeFlowOutputResponse_flowArn :: Lens' RemoveFlowOutputResponse (Maybe Text)

The ARN of the flow that is associated with the output you removed.

removeFlowOutputResponse_outputArn :: Lens' RemoveFlowOutputResponse (Maybe Text)

The ARN of the output that was removed.

removeFlowOutputResponse_httpStatus :: Lens' RemoveFlowOutputResponse Int

The response's http status code.