Amazonka.MediaConnect.RevokeFlowEntitlement

Description

Revokes an entitlement from a flow. Once an entitlement is revoked, the content becomes unavailable to the subscriber and the associated output is removed.

Creating a Request

data RevokeFlowEntitlement

See: newRevokeFlowEntitlement smart constructor.

Constructors

RevokeFlowEntitlement'

Fields

flowArn :: Text

The flow that you want to revoke an entitlement from.

entitlementArn :: Text

The ARN of the entitlement that you want to revoke.

newRevokeFlowEntitlement

Arguments

:: Text
$sel:flowArn:RevokeFlowEntitlement'

-> Text
$sel:entitlementArn:RevokeFlowEntitlement'

-> RevokeFlowEntitlement

Create a value of RevokeFlowEntitlement with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:flowArn:RevokeFlowEntitlement', revokeFlowEntitlement_flowArn - The flow that you want to revoke an entitlement from.

$sel:entitlementArn:RevokeFlowEntitlement', revokeFlowEntitlement_entitlementArn - The ARN of the entitlement that you want to revoke.

Request Lenses

revokeFlowEntitlement_flowArn :: Lens' RevokeFlowEntitlement Text

The flow that you want to revoke an entitlement from.

revokeFlowEntitlement_entitlementArn :: Lens' RevokeFlowEntitlement Text

The ARN of the entitlement that you want to revoke.

Destructuring the Response

data RevokeFlowEntitlementResponse

See: newRevokeFlowEntitlementResponse smart constructor.

Constructors

RevokeFlowEntitlementResponse'

Fields

entitlementArn :: Maybe Text

The ARN of the entitlement that was revoked.

flowArn :: Maybe Text

The ARN of the flow that the entitlement was revoked from.

httpStatus :: Int

The response's http status code.

newRevokeFlowEntitlementResponse

Arguments

:: Int
$sel:httpStatus:RevokeFlowEntitlementResponse'

-> RevokeFlowEntitlementResponse

Create a value of RevokeFlowEntitlementResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:entitlementArn:RevokeFlowEntitlement', revokeFlowEntitlementResponse_entitlementArn - The ARN of the entitlement that was revoked.

$sel:flowArn:RevokeFlowEntitlement', revokeFlowEntitlementResponse_flowArn - The ARN of the flow that the entitlement was revoked from.

$sel:httpStatus:RevokeFlowEntitlementResponse', revokeFlowEntitlementResponse_httpStatus - The response's http status code.

Response Lenses

revokeFlowEntitlementResponse_entitlementArn :: Lens' RevokeFlowEntitlementResponse (Maybe Text)

The ARN of the entitlement that was revoked.

revokeFlowEntitlementResponse_flowArn :: Lens' RevokeFlowEntitlementResponse (Maybe Text)

The ARN of the flow that the entitlement was revoked from.

revokeFlowEntitlementResponse_httpStatus :: Lens' RevokeFlowEntitlementResponse Int

The response's http status code.