Amazonka.MediaConnect.Types.AddOutputRequest

Description

Documentation

data AddOutputRequest

The output that you want to add to this flow.

See: newAddOutputRequest smart constructor.

Constructors

AddOutputRequest'

Fields

destination :: Maybe Text

The IP address from which video will be sent to output destinations.

maxLatency :: Maybe Int

The maximum latency in milliseconds for Zixi-based streams.

mediaStreamOutputConfigurations :: Maybe [MediaStreamOutputConfigurationRequest]

The media streams that are associated with the output, and the parameters for those associations.

encryption :: Maybe Encryption

The type of key used for the encryption. If no keyType is provided, the service will use the default setting (static-key).

name :: Maybe Text

The name of the output. This value must be unique within the current flow.

cidrAllowList :: Maybe [Text]

The range of IP addresses that should be allowed to initiate output requests to this flow. These IP addresses should be in the form of a Classless Inter-Domain Routing (CIDR) block; for example, 10.0.0.0/16.

smoothingLatency :: Maybe Int

The smoothing latency in milliseconds for RIST, RTP, and RTP-FEC streams.

minLatency :: Maybe Int

The minimum latency in milliseconds for SRT-based streams. In streams that use the SRT protocol, this value that you set on your MediaConnect source or output represents the minimal potential latency of that connection. The latency of the stream is set to the highest number between the sender’s minimum latency and the receiver’s minimum latency.

description :: Maybe Text

A description of the output. This description appears only on the AWS Elemental MediaConnect console and will not be seen by the end user.

port :: Maybe Int

The port to use when content is distributed to this output.

streamId :: Maybe Text

The stream ID that you want to use for this transport. This parameter applies only to Zixi-based streams.

remoteId :: Maybe Text

The remote ID for the Zixi-pull output stream.

vpcInterfaceAttachment :: Maybe VpcInterfaceAttachment

The name of the VPC interface attachment to use for this output.

protocol :: Protocol

The protocol to use for the output.

newAddOutputRequest

Arguments

:: Protocol
$sel:protocol:AddOutputRequest'

-> AddOutputRequest

Create a value of AddOutputRequest with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:destination:AddOutputRequest', addOutputRequest_destination - The IP address from which video will be sent to output destinations.

$sel:maxLatency:AddOutputRequest', addOutputRequest_maxLatency - The maximum latency in milliseconds for Zixi-based streams.

$sel:mediaStreamOutputConfigurations:AddOutputRequest', addOutputRequest_mediaStreamOutputConfigurations - The media streams that are associated with the output, and the parameters for those associations.

$sel:encryption:AddOutputRequest', addOutputRequest_encryption - The type of key used for the encryption. If no keyType is provided, the service will use the default setting (static-key).

$sel:name:AddOutputRequest', addOutputRequest_name - The name of the output. This value must be unique within the current flow.

$sel:cidrAllowList:AddOutputRequest', addOutputRequest_cidrAllowList - The range of IP addresses that should be allowed to initiate output requests to this flow. These IP addresses should be in the form of a Classless Inter-Domain Routing (CIDR) block; for example, 10.0.0.0/16.

$sel:smoothingLatency:AddOutputRequest', addOutputRequest_smoothingLatency - The smoothing latency in milliseconds for RIST, RTP, and RTP-FEC streams.

$sel:minLatency:AddOutputRequest', addOutputRequest_minLatency - The minimum latency in milliseconds for SRT-based streams. In streams that use the SRT protocol, this value that you set on your MediaConnect source or output represents the minimal potential latency of that connection. The latency of the stream is set to the highest number between the sender’s minimum latency and the receiver’s minimum latency.

$sel:description:AddOutputRequest', addOutputRequest_description - A description of the output. This description appears only on the AWS Elemental MediaConnect console and will not be seen by the end user.

$sel:port:AddOutputRequest', addOutputRequest_port - The port to use when content is distributed to this output.

$sel:streamId:AddOutputRequest', addOutputRequest_streamId - The stream ID that you want to use for this transport. This parameter applies only to Zixi-based streams.

$sel:remoteId:AddOutputRequest', addOutputRequest_remoteId - The remote ID for the Zixi-pull output stream.

$sel:vpcInterfaceAttachment:AddOutputRequest', addOutputRequest_vpcInterfaceAttachment - The name of the VPC interface attachment to use for this output.

$sel:protocol:AddOutputRequest', addOutputRequest_protocol - The protocol to use for the output.

addOutputRequest_destination :: Lens' AddOutputRequest (Maybe Text)

The IP address from which video will be sent to output destinations.

addOutputRequest_maxLatency :: Lens' AddOutputRequest (Maybe Int)

The maximum latency in milliseconds for Zixi-based streams.

addOutputRequest_mediaStreamOutputConfigurations :: Lens' AddOutputRequest (Maybe [MediaStreamOutputConfigurationRequest])

The media streams that are associated with the output, and the parameters for those associations.

addOutputRequest_encryption :: Lens' AddOutputRequest (Maybe Encryption)

The type of key used for the encryption. If no keyType is provided, the service will use the default setting (static-key).

addOutputRequest_name :: Lens' AddOutputRequest (Maybe Text)

The name of the output. This value must be unique within the current flow.

addOutputRequest_cidrAllowList :: Lens' AddOutputRequest (Maybe [Text])

The range of IP addresses that should be allowed to initiate output requests to this flow. These IP addresses should be in the form of a Classless Inter-Domain Routing (CIDR) block; for example, 10.0.0.0/16.

addOutputRequest_smoothingLatency :: Lens' AddOutputRequest (Maybe Int)

The smoothing latency in milliseconds for RIST, RTP, and RTP-FEC streams.

addOutputRequest_minLatency :: Lens' AddOutputRequest (Maybe Int)

The minimum latency in milliseconds for SRT-based streams. In streams that use the SRT protocol, this value that you set on your MediaConnect source or output represents the minimal potential latency of that connection. The latency of the stream is set to the highest number between the sender’s minimum latency and the receiver’s minimum latency.

addOutputRequest_description :: Lens' AddOutputRequest (Maybe Text)

A description of the output. This description appears only on the AWS Elemental MediaConnect console and will not be seen by the end user.

addOutputRequest_port :: Lens' AddOutputRequest (Maybe Int)

The port to use when content is distributed to this output.

addOutputRequest_streamId :: Lens' AddOutputRequest (Maybe Text)

The stream ID that you want to use for this transport. This parameter applies only to Zixi-based streams.

addOutputRequest_remoteId :: Lens' AddOutputRequest (Maybe Text)

The remote ID for the Zixi-pull output stream.

addOutputRequest_vpcInterfaceAttachment :: Lens' AddOutputRequest (Maybe VpcInterfaceAttachment)

The name of the VPC interface attachment to use for this output.

addOutputRequest_protocol :: Lens' AddOutputRequest Protocol

The protocol to use for the output.