Amazonka.MediaConnect.Types.DestinationConfiguration

Description

Documentation

data DestinationConfiguration

The transport parameters that are associated with an outbound media stream.

See: newDestinationConfiguration smart constructor.

Constructors

DestinationConfiguration'

Fields

destinationIp :: Text

The IP address where contents of the media stream will be sent.

destinationPort :: Int

The port to use when the content of the media stream is distributed to the output.

interface :: Interface

The VPC interface that is used for the media stream associated with the output.

outboundIp :: Text

The IP address that the receiver requires in order to establish a connection with the flow. This value is represented by the elastic network interface IP address of the VPC. This field applies only to outputs that use the CDI or ST 2110 JPEG XS protocol.

newDestinationConfiguration

Arguments

:: Text
$sel:destinationIp:DestinationConfiguration'

-> Int
$sel:destinationPort:DestinationConfiguration'

-> Interface
$sel:interface:DestinationConfiguration'

-> Text
$sel:outboundIp:DestinationConfiguration'

-> DestinationConfiguration

Create a value of DestinationConfiguration with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:destinationIp:DestinationConfiguration', destinationConfiguration_destinationIp - The IP address where contents of the media stream will be sent.

$sel:destinationPort:DestinationConfiguration', destinationConfiguration_destinationPort - The port to use when the content of the media stream is distributed to the output.

$sel:interface:DestinationConfiguration', destinationConfiguration_interface - The VPC interface that is used for the media stream associated with the output.

$sel:outboundIp:DestinationConfiguration', destinationConfiguration_outboundIp - The IP address that the receiver requires in order to establish a connection with the flow. This value is represented by the elastic network interface IP address of the VPC. This field applies only to outputs that use the CDI or ST 2110 JPEG XS protocol.

destinationConfiguration_destinationIp :: Lens' DestinationConfiguration Text

The IP address where contents of the media stream will be sent.

destinationConfiguration_destinationPort :: Lens' DestinationConfiguration Int

The port to use when the content of the media stream is distributed to the output.

destinationConfiguration_interface :: Lens' DestinationConfiguration Interface

The VPC interface that is used for the media stream associated with the output.

destinationConfiguration_outboundIp :: Lens' DestinationConfiguration Text

The IP address that the receiver requires in order to establish a connection with the flow. This value is represented by the elastic network interface IP address of the VPC. This field applies only to outputs that use the CDI or ST 2110 JPEG XS protocol.