Amazonka.MediaConnect.Types.EncodingParameters

Description

Documentation

data EncodingParameters

A collection of parameters that determine how MediaConnect will convert the content. These fields only apply to outputs on flows that have a CDI source.

See: newEncodingParameters smart constructor.

Constructors

EncodingParameters'

Fields

encoderProfile :: EncoderProfile

A setting on the encoder that drives compression settings. This property only applies to video media streams associated with outputs that use the ST 2110 JPEG XS protocol, with a flow source that uses the CDI protocol.

compressionFactor :: Double

A value that is used to calculate compression for an output. The bitrate of the output is calculated as follows: Output bitrate = (1 / compressionFactor) * (source bitrate) This property only applies to outputs that use the ST 2110 JPEG XS protocol, with a flow source that uses the CDI protocol. Valid values are floating point numbers in the range of 3.0 to 10.0, inclusive.

newEncodingParameters

Arguments

:: EncoderProfile
$sel:encoderProfile:EncodingParameters'

-> Double
$sel:compressionFactor:EncodingParameters'

-> EncodingParameters

Create a value of EncodingParameters with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:encoderProfile:EncodingParameters', encodingParameters_encoderProfile - A setting on the encoder that drives compression settings. This property only applies to video media streams associated with outputs that use the ST 2110 JPEG XS protocol, with a flow source that uses the CDI protocol.

$sel:compressionFactor:EncodingParameters', encodingParameters_compressionFactor - A value that is used to calculate compression for an output. The bitrate of the output is calculated as follows: Output bitrate = (1 / compressionFactor) * (source bitrate) This property only applies to outputs that use the ST 2110 JPEG XS protocol, with a flow source that uses the CDI protocol. Valid values are floating point numbers in the range of 3.0 to 10.0, inclusive.

encodingParameters_encoderProfile :: Lens' EncodingParameters EncoderProfile

A setting on the encoder that drives compression settings. This property only applies to video media streams associated with outputs that use the ST 2110 JPEG XS protocol, with a flow source that uses the CDI protocol.

encodingParameters_compressionFactor :: Lens' EncodingParameters Double

A value that is used to calculate compression for an output. The bitrate of the output is calculated as follows: Output bitrate = (1 / compressionFactor) * (source bitrate) This property only applies to outputs that use the ST 2110 JPEG XS protocol, with a flow source that uses the CDI protocol. Valid values are floating point numbers in the range of 3.0 to 10.0, inclusive.