Amazonka.MediaConnect.Types.Flow

Description

Documentation

data Flow

The settings for a flow, including its source, outputs, and entitlements.

See: newFlow smart constructor.

Constructors

Flow'

Fields

mediaStreams :: Maybe [MediaStream]

The media streams that are associated with the flow. After you associate a media stream with a source, you can also associate it with outputs on the flow.

sourceFailoverConfig :: Maybe FailoverConfig

vpcInterfaces :: Maybe [VpcInterface]

The VPC Interfaces for this flow.

sources :: Maybe [Source]

egressIp :: Maybe Text

The IP address from which video will be sent to output destinations.

description :: Maybe Text

A description of the flow. This value is not used or seen outside of the current AWS Elemental MediaConnect account.

status :: Status

The current status of the flow.

entitlements :: [Entitlement]

The entitlements in this flow.

outputs :: [Output]

The outputs in this flow.

availabilityZone :: Text

The Availability Zone that you want to create the flow in. These options are limited to the Availability Zones within the current AWS.

flowArn :: Text

The Amazon Resource Name (ARN), a unique identifier for any AWS resource, of the flow.

source :: Source

name :: Text

The name of the flow.

newFlow

Arguments

:: Status
$sel:status:Flow'

-> Text
$sel:availabilityZone:Flow'

-> Text
$sel:flowArn:Flow'

-> Source
$sel:source:Flow'

-> Text
$sel:name:Flow'

-> Flow

Create a value of Flow with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:mediaStreams:Flow', flow_mediaStreams - The media streams that are associated with the flow. After you associate a media stream with a source, you can also associate it with outputs on the flow.

$sel:sourceFailoverConfig:Flow', flow_sourceFailoverConfig - Undocumented member.

$sel:vpcInterfaces:Flow', flow_vpcInterfaces - The VPC Interfaces for this flow.

$sel:sources:Flow', flow_sources - Undocumented member.

$sel:egressIp:Flow', flow_egressIp - The IP address from which video will be sent to output destinations.

$sel:description:Flow', flow_description - A description of the flow. This value is not used or seen outside of the current AWS Elemental MediaConnect account.

$sel:status:Flow', flow_status - The current status of the flow.

$sel:entitlements:Flow', flow_entitlements - The entitlements in this flow.

$sel:outputs:Flow', flow_outputs - The outputs in this flow.

$sel:availabilityZone:Flow', flow_availabilityZone - The Availability Zone that you want to create the flow in. These options are limited to the Availability Zones within the current AWS.

$sel:flowArn:Flow', flow_flowArn - The Amazon Resource Name (ARN), a unique identifier for any AWS resource, of the flow.

$sel:source:Flow', flow_source - Undocumented member.

$sel:name:Flow', flow_name - The name of the flow.

flow_mediaStreams :: Lens' Flow (Maybe [MediaStream])

The media streams that are associated with the flow. After you associate a media stream with a source, you can also associate it with outputs on the flow.

flow_sourceFailoverConfig :: Lens' Flow (Maybe FailoverConfig)

Undocumented member.

flow_vpcInterfaces :: Lens' Flow (Maybe [VpcInterface])

The VPC Interfaces for this flow.

flow_sources :: Lens' Flow (Maybe [Source])

Undocumented member.

flow_egressIp :: Lens' Flow (Maybe Text)

The IP address from which video will be sent to output destinations.

flow_description :: Lens' Flow (Maybe Text)

A description of the flow. This value is not used or seen outside of the current AWS Elemental MediaConnect account.

flow_status :: Lens' Flow Status

The current status of the flow.

flow_entitlements :: Lens' Flow [Entitlement]

The entitlements in this flow.

flow_outputs :: Lens' Flow [Output]

The outputs in this flow.

flow_availabilityZone :: Lens' Flow Text

The Availability Zone that you want to create the flow in. These options are limited to the Availability Zones within the current AWS.

flow_flowArn :: Lens' Flow Text

The Amazon Resource Name (ARN), a unique identifier for any AWS resource, of the flow.

flow_source :: Lens' Flow Source

Undocumented member.

flow_name :: Lens' Flow Text

The name of the flow.