Amazonka.MediaConnect.Types.InputConfigurationRequest

Description

Documentation

data InputConfigurationRequest

The transport parameters that you want to associate with an incoming media stream.

See: newInputConfigurationRequest smart constructor.

Constructors

InputConfigurationRequest'

Fields

inputPort :: Int

The port that you want the flow to listen on for an incoming media stream.

interface :: InterfaceRequest

The VPC interface that you want to use for the incoming media stream.

newInputConfigurationRequest

Arguments

:: Int
$sel:inputPort:InputConfigurationRequest'

-> InterfaceRequest
$sel:interface:InputConfigurationRequest'

-> InputConfigurationRequest

Create a value of InputConfigurationRequest with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:inputPort:InputConfigurationRequest', inputConfigurationRequest_inputPort - The port that you want the flow to listen on for an incoming media stream.

$sel:interface:InputConfigurationRequest', inputConfigurationRequest_interface - The VPC interface that you want to use for the incoming media stream.

inputConfigurationRequest_inputPort :: Lens' InputConfigurationRequest Int

The port that you want the flow to listen on for an incoming media stream.

inputConfigurationRequest_interface :: Lens' InputConfigurationRequest InterfaceRequest

The VPC interface that you want to use for the incoming media stream.