Amazonka.MediaConnect.Types.ListedFlow

Description

Documentation

data ListedFlow

Provides a summary of a flow, including its ARN, Availability Zone, and source type.

See: newListedFlow smart constructor.

Constructors

ListedFlow'

Fields

status :: Status

The current status of the flow.

description :: Text

A description of the flow.

sourceType :: SourceType

The type of source. This value is either owned (originated somewhere other than an AWS Elemental MediaConnect flow owned by another AWS account) or entitled (originated at an AWS Elemental MediaConnect flow owned by another AWS account).

availabilityZone :: Text

The Availability Zone that the flow was created in.

flowArn :: Text

The ARN of the flow.

name :: Text

The name of the flow.

newListedFlow

Arguments

:: Status
$sel:status:ListedFlow'

-> Text
$sel:description:ListedFlow'

-> SourceType
$sel:sourceType:ListedFlow'

-> Text
$sel:availabilityZone:ListedFlow'

-> Text
$sel:flowArn:ListedFlow'

-> Text
$sel:name:ListedFlow'

-> ListedFlow

Create a value of ListedFlow with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:status:ListedFlow', listedFlow_status - The current status of the flow.

$sel:description:ListedFlow', listedFlow_description - A description of the flow.

$sel:sourceType:ListedFlow', listedFlow_sourceType - The type of source. This value is either owned (originated somewhere other than an AWS Elemental MediaConnect flow owned by another AWS account) or entitled (originated at an AWS Elemental MediaConnect flow owned by another AWS account).

$sel:availabilityZone:ListedFlow', listedFlow_availabilityZone - The Availability Zone that the flow was created in.

$sel:flowArn:ListedFlow', listedFlow_flowArn - The ARN of the flow.

$sel:name:ListedFlow', listedFlow_name - The name of the flow.

listedFlow_status :: Lens' ListedFlow Status

The current status of the flow.

listedFlow_description :: Lens' ListedFlow Text

A description of the flow.

listedFlow_sourceType :: Lens' ListedFlow SourceType

The type of source. This value is either owned (originated somewhere other than an AWS Elemental MediaConnect flow owned by another AWS account) or entitled (originated at an AWS Elemental MediaConnect flow owned by another AWS account).

listedFlow_availabilityZone :: Lens' ListedFlow Text

The Availability Zone that the flow was created in.

listedFlow_flowArn :: Lens' ListedFlow Text

The ARN of the flow.

listedFlow_name :: Lens' ListedFlow Text

The name of the flow.