Amazonka.MediaConnect.Types.MediaStreamSourceConfiguration

Description

Documentation

data MediaStreamSourceConfiguration

The media stream that is associated with the source, and the parameters for that association.

See: newMediaStreamSourceConfiguration smart constructor.

Constructors

MediaStreamSourceConfiguration'

Fields

inputConfigurations :: Maybe [InputConfiguration]

The transport parameters that are associated with an incoming media stream.

mediaStreamName :: Text

The name of the media stream.

encodingName :: EncodingName

The format that was used to encode the data. For ancillary data streams, set the encoding name to smpte291. For audio streams, set the encoding name to pcm. For video, 2110 streams, set the encoding name to raw. For video, JPEG XS streams, set the encoding name to jxsv.

newMediaStreamSourceConfiguration

Arguments

:: Text
$sel:mediaStreamName:MediaStreamSourceConfiguration'

-> EncodingName
$sel:encodingName:MediaStreamSourceConfiguration'

-> MediaStreamSourceConfiguration

Create a value of MediaStreamSourceConfiguration with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:inputConfigurations:MediaStreamSourceConfiguration', mediaStreamSourceConfiguration_inputConfigurations - The transport parameters that are associated with an incoming media stream.

$sel:mediaStreamName:MediaStreamSourceConfiguration', mediaStreamSourceConfiguration_mediaStreamName - The name of the media stream.

$sel:encodingName:MediaStreamSourceConfiguration', mediaStreamSourceConfiguration_encodingName - The format that was used to encode the data. For ancillary data streams, set the encoding name to smpte291. For audio streams, set the encoding name to pcm. For video, 2110 streams, set the encoding name to raw. For video, JPEG XS streams, set the encoding name to jxsv.

mediaStreamSourceConfiguration_inputConfigurations :: Lens' MediaStreamSourceConfiguration (Maybe [InputConfiguration])

The transport parameters that are associated with an incoming media stream.

mediaStreamSourceConfiguration_mediaStreamName :: Lens' MediaStreamSourceConfiguration Text

The name of the media stream.

mediaStreamSourceConfiguration_encodingName :: Lens' MediaStreamSourceConfiguration EncodingName

The format that was used to encode the data. For ancillary data streams, set the encoding name to smpte291. For audio streams, set the encoding name to pcm. For video, 2110 streams, set the encoding name to raw. For video, JPEG XS streams, set the encoding name to jxsv.