Amazonka.MediaConnect.Types.MediaStreamSourceConfigurationRequest

Description

Documentation

data MediaStreamSourceConfigurationRequest

The definition of a media stream that you want to associate with the source.

See: newMediaStreamSourceConfigurationRequest smart constructor.

Constructors

MediaStreamSourceConfigurationRequest'

Fields

inputConfigurations :: Maybe [InputConfigurationRequest]

The transport parameters that you want to associate with the media stream.

mediaStreamName :: Text

The name of the media stream.

encodingName :: EncodingName

The format you want to use to encode the data. For ancillary data streams, set the encoding name to smpte291. For audio streams, set the encoding name to pcm. For video, 2110 streams, set the encoding name to raw. For video, JPEG XS streams, set the encoding name to jxsv.

newMediaStreamSourceConfigurationRequest

Arguments

:: Text
$sel:mediaStreamName:MediaStreamSourceConfigurationRequest'

-> EncodingName
$sel:encodingName:MediaStreamSourceConfigurationRequest'

-> MediaStreamSourceConfigurationRequest

Create a value of MediaStreamSourceConfigurationRequest with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:inputConfigurations:MediaStreamSourceConfigurationRequest', mediaStreamSourceConfigurationRequest_inputConfigurations - The transport parameters that you want to associate with the media stream.

$sel:mediaStreamName:MediaStreamSourceConfigurationRequest', mediaStreamSourceConfigurationRequest_mediaStreamName - The name of the media stream.

$sel:encodingName:MediaStreamSourceConfigurationRequest', mediaStreamSourceConfigurationRequest_encodingName - The format you want to use to encode the data. For ancillary data streams, set the encoding name to smpte291. For audio streams, set the encoding name to pcm. For video, 2110 streams, set the encoding name to raw. For video, JPEG XS streams, set the encoding name to jxsv.

mediaStreamSourceConfigurationRequest_inputConfigurations :: Lens' MediaStreamSourceConfigurationRequest (Maybe [InputConfigurationRequest])

The transport parameters that you want to associate with the media stream.

mediaStreamSourceConfigurationRequest_mediaStreamName :: Lens' MediaStreamSourceConfigurationRequest Text

The name of the media stream.

mediaStreamSourceConfigurationRequest_encodingName :: Lens' MediaStreamSourceConfigurationRequest EncodingName

The format you want to use to encode the data. For ancillary data streams, set the encoding name to smpte291. For audio streams, set the encoding name to pcm. For video, 2110 streams, set the encoding name to raw. For video, JPEG XS streams, set the encoding name to jxsv.