Amazonka.MediaConnect.Types.SetSourceRequest

Description

Documentation

data SetSourceRequest

The settings for the source of the flow.

See: newSetSourceRequest smart constructor.

Constructors

SetSourceRequest'

Fields

entitlementArn :: Maybe Text

The ARN of the entitlement that allows you to subscribe to this flow. The entitlement is set by the flow originator, and the ARN is generated as part of the originator's flow.

maxLatency :: Maybe Int

The maximum latency in milliseconds. This parameter applies only to RIST-based and Zixi-based streams.

vpcInterfaceName :: Maybe Text

The name of the VPC interface to use for this source.

decryption :: Maybe Encryption

The type of encryption that is used on the content ingested from this source.

maxSyncBuffer :: Maybe Int

The size of the buffer (in milliseconds) to use to sync incoming source data.

protocol :: Maybe Protocol

The protocol that is used by the source.

name :: Maybe Text

The name of the source.

minLatency :: Maybe Int

The minimum latency in milliseconds for SRT-based streams. In streams that use the SRT protocol, this value that you set on your MediaConnect source or output represents the minimal potential latency of that connection. The latency of the stream is set to the highest number between the sender’s minimum latency and the receiver’s minimum latency.

ingestPort :: Maybe Int

The port that the flow will be listening on for incoming content.

description :: Maybe Text

A description for the source. This value is not used or seen outside of the current AWS Elemental MediaConnect account.

mediaStreamSourceConfigurations :: Maybe [MediaStreamSourceConfigurationRequest]

The media streams that are associated with the source, and the parameters for those associations.

whitelistCidr :: Maybe Text

The range of IP addresses that should be allowed to contribute content to your source. These IP addresses should be in the form of a Classless Inter-Domain Routing (CIDR) block; for example, 10.0.0.0/16.

maxBitrate :: Maybe Int

The smoothing max bitrate for RIST, RTP, and RTP-FEC streams.

streamId :: Maybe Text

The stream ID that you want to use for this transport. This parameter applies only to Zixi-based streams.

newSetSourceRequest :: SetSourceRequest

Create a value of SetSourceRequest with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:entitlementArn:SetSourceRequest', setSourceRequest_entitlementArn - The ARN of the entitlement that allows you to subscribe to this flow. The entitlement is set by the flow originator, and the ARN is generated as part of the originator's flow.

$sel:maxLatency:SetSourceRequest', setSourceRequest_maxLatency - The maximum latency in milliseconds. This parameter applies only to RIST-based and Zixi-based streams.

$sel:vpcInterfaceName:SetSourceRequest', setSourceRequest_vpcInterfaceName - The name of the VPC interface to use for this source.

$sel:decryption:SetSourceRequest', setSourceRequest_decryption - The type of encryption that is used on the content ingested from this source.

$sel:maxSyncBuffer:SetSourceRequest', setSourceRequest_maxSyncBuffer - The size of the buffer (in milliseconds) to use to sync incoming source data.

$sel:protocol:SetSourceRequest', setSourceRequest_protocol - The protocol that is used by the source.

$sel:name:SetSourceRequest', setSourceRequest_name - The name of the source.

$sel:minLatency:SetSourceRequest', setSourceRequest_minLatency - The minimum latency in milliseconds for SRT-based streams. In streams that use the SRT protocol, this value that you set on your MediaConnect source or output represents the minimal potential latency of that connection. The latency of the stream is set to the highest number between the sender’s minimum latency and the receiver’s minimum latency.

$sel:ingestPort:SetSourceRequest', setSourceRequest_ingestPort - The port that the flow will be listening on for incoming content.

$sel:description:SetSourceRequest', setSourceRequest_description - A description for the source. This value is not used or seen outside of the current AWS Elemental MediaConnect account.

$sel:mediaStreamSourceConfigurations:SetSourceRequest', setSourceRequest_mediaStreamSourceConfigurations - The media streams that are associated with the source, and the parameters for those associations.

$sel:whitelistCidr:SetSourceRequest', setSourceRequest_whitelistCidr - The range of IP addresses that should be allowed to contribute content to your source. These IP addresses should be in the form of a Classless Inter-Domain Routing (CIDR) block; for example, 10.0.0.0/16.

$sel:maxBitrate:SetSourceRequest', setSourceRequest_maxBitrate - The smoothing max bitrate for RIST, RTP, and RTP-FEC streams.

$sel:streamId:SetSourceRequest', setSourceRequest_streamId - The stream ID that you want to use for this transport. This parameter applies only to Zixi-based streams.

setSourceRequest_entitlementArn :: Lens' SetSourceRequest (Maybe Text)

The ARN of the entitlement that allows you to subscribe to this flow. The entitlement is set by the flow originator, and the ARN is generated as part of the originator's flow.

setSourceRequest_maxLatency :: Lens' SetSourceRequest (Maybe Int)

The maximum latency in milliseconds. This parameter applies only to RIST-based and Zixi-based streams.

setSourceRequest_vpcInterfaceName :: Lens' SetSourceRequest (Maybe Text)

The name of the VPC interface to use for this source.

setSourceRequest_decryption :: Lens' SetSourceRequest (Maybe Encryption)

The type of encryption that is used on the content ingested from this source.

setSourceRequest_maxSyncBuffer :: Lens' SetSourceRequest (Maybe Int)

The size of the buffer (in milliseconds) to use to sync incoming source data.

setSourceRequest_protocol :: Lens' SetSourceRequest (Maybe Protocol)

The protocol that is used by the source.

setSourceRequest_name :: Lens' SetSourceRequest (Maybe Text)

The name of the source.

setSourceRequest_minLatency :: Lens' SetSourceRequest (Maybe Int)

The minimum latency in milliseconds for SRT-based streams. In streams that use the SRT protocol, this value that you set on your MediaConnect source or output represents the minimal potential latency of that connection. The latency of the stream is set to the highest number between the sender’s minimum latency and the receiver’s minimum latency.

setSourceRequest_ingestPort :: Lens' SetSourceRequest (Maybe Int)

The port that the flow will be listening on for incoming content.

setSourceRequest_description :: Lens' SetSourceRequest (Maybe Text)

A description for the source. This value is not used or seen outside of the current AWS Elemental MediaConnect account.

setSourceRequest_mediaStreamSourceConfigurations :: Lens' SetSourceRequest (Maybe [MediaStreamSourceConfigurationRequest])

The media streams that are associated with the source, and the parameters for those associations.

setSourceRequest_whitelistCidr :: Lens' SetSourceRequest (Maybe Text)

The range of IP addresses that should be allowed to contribute content to your source. These IP addresses should be in the form of a Classless Inter-Domain Routing (CIDR) block; for example, 10.0.0.0/16.

setSourceRequest_maxBitrate :: Lens' SetSourceRequest (Maybe Int)

The smoothing max bitrate for RIST, RTP, and RTP-FEC streams.

setSourceRequest_streamId :: Lens' SetSourceRequest (Maybe Text)

The stream ID that you want to use for this transport. This parameter applies only to Zixi-based streams.