Amazonka.MediaConnect.Types.Transport

Description

Documentation

data Transport

Attributes related to the transport stream that are used in a source or output.

See: newTransport smart constructor.

Constructors

Transport'

Fields

maxLatency :: Maybe Int

The maximum latency in milliseconds. This parameter applies only to RIST-based and Zixi-based streams.

maxSyncBuffer :: Maybe Int

The size of the buffer (in milliseconds) to use to sync incoming source data.

cidrAllowList :: Maybe [Text]

The range of IP addresses that should be allowed to initiate output requests to this flow. These IP addresses should be in the form of a Classless Inter-Domain Routing (CIDR) block; for example, 10.0.0.0/16.

smoothingLatency :: Maybe Int

The smoothing latency in milliseconds for RIST, RTP, and RTP-FEC streams.

minLatency :: Maybe Int

The minimum latency in milliseconds for SRT-based streams. In streams that use the SRT protocol, this value that you set on your MediaConnect source or output represents the minimal potential latency of that connection. The latency of the stream is set to the highest number between the sender’s minimum latency and the receiver’s minimum latency.

maxBitrate :: Maybe Int

The smoothing max bitrate for RIST, RTP, and RTP-FEC streams.

streamId :: Maybe Text

The stream ID that you want to use for this transport. This parameter applies only to Zixi-based streams.

remoteId :: Maybe Text

The remote ID for the Zixi-pull stream.

protocol :: Protocol

The protocol that is used by the source or output.

newTransport

Arguments

:: Protocol
$sel:protocol:Transport'

-> Transport

Create a value of Transport with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:maxLatency:Transport', transport_maxLatency - The maximum latency in milliseconds. This parameter applies only to RIST-based and Zixi-based streams.

$sel:maxSyncBuffer:Transport', transport_maxSyncBuffer - The size of the buffer (in milliseconds) to use to sync incoming source data.

$sel:cidrAllowList:Transport', transport_cidrAllowList - The range of IP addresses that should be allowed to initiate output requests to this flow. These IP addresses should be in the form of a Classless Inter-Domain Routing (CIDR) block; for example, 10.0.0.0/16.

$sel:smoothingLatency:Transport', transport_smoothingLatency - The smoothing latency in milliseconds for RIST, RTP, and RTP-FEC streams.

$sel:minLatency:Transport', transport_minLatency - The minimum latency in milliseconds for SRT-based streams. In streams that use the SRT protocol, this value that you set on your MediaConnect source or output represents the minimal potential latency of that connection. The latency of the stream is set to the highest number between the sender’s minimum latency and the receiver’s minimum latency.

$sel:maxBitrate:Transport', transport_maxBitrate - The smoothing max bitrate for RIST, RTP, and RTP-FEC streams.

$sel:streamId:Transport', transport_streamId - The stream ID that you want to use for this transport. This parameter applies only to Zixi-based streams.

$sel:remoteId:Transport', transport_remoteId - The remote ID for the Zixi-pull stream.

$sel:protocol:Transport', transport_protocol - The protocol that is used by the source or output.

transport_maxLatency :: Lens' Transport (Maybe Int)

The maximum latency in milliseconds. This parameter applies only to RIST-based and Zixi-based streams.

transport_maxSyncBuffer :: Lens' Transport (Maybe Int)

The size of the buffer (in milliseconds) to use to sync incoming source data.

transport_cidrAllowList :: Lens' Transport (Maybe [Text])

The range of IP addresses that should be allowed to initiate output requests to this flow. These IP addresses should be in the form of a Classless Inter-Domain Routing (CIDR) block; for example, 10.0.0.0/16.

transport_smoothingLatency :: Lens' Transport (Maybe Int)

The smoothing latency in milliseconds for RIST, RTP, and RTP-FEC streams.

transport_minLatency :: Lens' Transport (Maybe Int)

The minimum latency in milliseconds for SRT-based streams. In streams that use the SRT protocol, this value that you set on your MediaConnect source or output represents the minimal potential latency of that connection. The latency of the stream is set to the highest number between the sender’s minimum latency and the receiver’s minimum latency.

transport_maxBitrate :: Lens' Transport (Maybe Int)

The smoothing max bitrate for RIST, RTP, and RTP-FEC streams.

transport_streamId :: Lens' Transport (Maybe Text)

The stream ID that you want to use for this transport. This parameter applies only to Zixi-based streams.

transport_remoteId :: Lens' Transport (Maybe Text)

The remote ID for the Zixi-pull stream.

transport_protocol :: Lens' Transport Protocol

The protocol that is used by the source or output.