Amazonka.MediaConnect.Types.UpdateEncryption

Description

Documentation

data UpdateEncryption

Information about the encryption of the flow.

See: newUpdateEncryption smart constructor.

Constructors

UpdateEncryption'

Fields

keyType :: Maybe KeyType

The type of key that is used for the encryption. If no keyType is provided, the service will use the default setting (static-key).

resourceId :: Maybe Text

An identifier for the content. The service sends this value to the key server to identify the current endpoint. The resource ID is also known as the content ID. This parameter is required for SPEKE encryption and is not valid for static key encryption.

url :: Maybe Text

The URL from the API Gateway proxy that you set up to talk to your key server. This parameter is required for SPEKE encryption and is not valid for static key encryption.

algorithm :: Maybe Algorithm

The type of algorithm that is used for the encryption (such as aes128, aes192, or aes256).

constantInitializationVector :: Maybe Text

A 128-bit, 16-byte hex value represented by a 32-character string, to be used with the key for encrypting content. This parameter is not valid for static key encryption.

deviceId :: Maybe Text

The value of one of the devices that you configured with your digital rights management (DRM) platform key provider. This parameter is required for SPEKE encryption and is not valid for static key encryption.

region :: Maybe Text

The AWS Region that the API Gateway proxy endpoint was created in. This parameter is required for SPEKE encryption and is not valid for static key encryption.

secretArn :: Maybe Text

The ARN of the secret that you created in AWS Secrets Manager to store the encryption key. This parameter is required for static key encryption and is not valid for SPEKE encryption.

roleArn :: Maybe Text

The ARN of the role that you created during setup (when you set up AWS Elemental MediaConnect as a trusted entity).

newUpdateEncryption :: UpdateEncryption

Create a value of UpdateEncryption with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:keyType:UpdateEncryption', updateEncryption_keyType - The type of key that is used for the encryption. If no keyType is provided, the service will use the default setting (static-key).

$sel:resourceId:UpdateEncryption', updateEncryption_resourceId - An identifier for the content. The service sends this value to the key server to identify the current endpoint. The resource ID is also known as the content ID. This parameter is required for SPEKE encryption and is not valid for static key encryption.

$sel:url:UpdateEncryption', updateEncryption_url - The URL from the API Gateway proxy that you set up to talk to your key server. This parameter is required for SPEKE encryption and is not valid for static key encryption.

$sel:algorithm:UpdateEncryption', updateEncryption_algorithm - The type of algorithm that is used for the encryption (such as aes128, aes192, or aes256).

$sel:constantInitializationVector:UpdateEncryption', updateEncryption_constantInitializationVector - A 128-bit, 16-byte hex value represented by a 32-character string, to be used with the key for encrypting content. This parameter is not valid for static key encryption.

$sel:deviceId:UpdateEncryption', updateEncryption_deviceId - The value of one of the devices that you configured with your digital rights management (DRM) platform key provider. This parameter is required for SPEKE encryption and is not valid for static key encryption.

$sel:region:UpdateEncryption', updateEncryption_region - The AWS Region that the API Gateway proxy endpoint was created in. This parameter is required for SPEKE encryption and is not valid for static key encryption.

$sel:secretArn:UpdateEncryption', updateEncryption_secretArn - The ARN of the secret that you created in AWS Secrets Manager to store the encryption key. This parameter is required for static key encryption and is not valid for SPEKE encryption.

$sel:roleArn:UpdateEncryption', updateEncryption_roleArn - The ARN of the role that you created during setup (when you set up AWS Elemental MediaConnect as a trusted entity).

updateEncryption_keyType :: Lens' UpdateEncryption (Maybe KeyType)

The type of key that is used for the encryption. If no keyType is provided, the service will use the default setting (static-key).

updateEncryption_resourceId :: Lens' UpdateEncryption (Maybe Text)

An identifier for the content. The service sends this value to the key server to identify the current endpoint. The resource ID is also known as the content ID. This parameter is required for SPEKE encryption and is not valid for static key encryption.

updateEncryption_url :: Lens' UpdateEncryption (Maybe Text)

The URL from the API Gateway proxy that you set up to talk to your key server. This parameter is required for SPEKE encryption and is not valid for static key encryption.

updateEncryption_algorithm :: Lens' UpdateEncryption (Maybe Algorithm)

The type of algorithm that is used for the encryption (such as aes128, aes192, or aes256).

updateEncryption_constantInitializationVector :: Lens' UpdateEncryption (Maybe Text)

A 128-bit, 16-byte hex value represented by a 32-character string, to be used with the key for encrypting content. This parameter is not valid for static key encryption.

updateEncryption_deviceId :: Lens' UpdateEncryption (Maybe Text)

The value of one of the devices that you configured with your digital rights management (DRM) platform key provider. This parameter is required for SPEKE encryption and is not valid for static key encryption.

updateEncryption_region :: Lens' UpdateEncryption (Maybe Text)

The AWS Region that the API Gateway proxy endpoint was created in. This parameter is required for SPEKE encryption and is not valid for static key encryption.

updateEncryption_secretArn :: Lens' UpdateEncryption (Maybe Text)

The ARN of the secret that you created in AWS Secrets Manager to store the encryption key. This parameter is required for static key encryption and is not valid for SPEKE encryption.

updateEncryption_roleArn :: Lens' UpdateEncryption (Maybe Text)

The ARN of the role that you created during setup (when you set up AWS Elemental MediaConnect as a trusted entity).