Amazonka.MediaConnect.Types.UpdateFailoverConfig

Description

Documentation

data UpdateFailoverConfig

The settings for source failover

See: newUpdateFailoverConfig smart constructor.

Constructors

UpdateFailoverConfig'

Fields

state :: Maybe State

recoveryWindow :: Maybe Int

Recovery window time to look for dash-7 packets

sourcePriority :: Maybe SourcePriority

The priority you want to assign to a source. You can have a primary stream and a backup stream or two equally prioritized streams.

failoverMode :: Maybe FailoverMode

The type of failover you choose for this flow. MERGE combines the source streams into a single stream, allowing graceful recovery from any single-source loss. FAILOVER allows switching between different streams.

newUpdateFailoverConfig :: UpdateFailoverConfig

Create a value of UpdateFailoverConfig with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:state:UpdateFailoverConfig', updateFailoverConfig_state - Undocumented member.

$sel:recoveryWindow:UpdateFailoverConfig', updateFailoverConfig_recoveryWindow - Recovery window time to look for dash-7 packets

$sel:sourcePriority:UpdateFailoverConfig', updateFailoverConfig_sourcePriority - The priority you want to assign to a source. You can have a primary stream and a backup stream or two equally prioritized streams.

$sel:failoverMode:UpdateFailoverConfig', updateFailoverConfig_failoverMode - The type of failover you choose for this flow. MERGE combines the source streams into a single stream, allowing graceful recovery from any single-source loss. FAILOVER allows switching between different streams.

updateFailoverConfig_state :: Lens' UpdateFailoverConfig (Maybe State)

Undocumented member.

updateFailoverConfig_recoveryWindow :: Lens' UpdateFailoverConfig (Maybe Int)

Recovery window time to look for dash-7 packets

updateFailoverConfig_sourcePriority :: Lens' UpdateFailoverConfig (Maybe SourcePriority)

The priority you want to assign to a source. You can have a primary stream and a backup stream or two equally prioritized streams.

updateFailoverConfig_failoverMode :: Lens' UpdateFailoverConfig (Maybe FailoverMode)

The type of failover you choose for this flow. MERGE combines the source streams into a single stream, allowing graceful recovery from any single-source loss. FAILOVER allows switching between different streams.