Amazonka.MediaConnect.Types.VpcInterfaceRequest

Description

Documentation

data VpcInterfaceRequest

Desired VPC Interface for a Flow

See: newVpcInterfaceRequest smart constructor.

Constructors

VpcInterfaceRequest'

Fields

networkInterfaceType :: Maybe NetworkInterfaceType

The type of network interface. If this value is not included in the request, MediaConnect uses ENA as the networkInterfaceType.

subnetId :: Text

Subnet must be in the AZ of the Flow

securityGroupIds :: [Text]

Security Group IDs to be used on ENI.

roleArn :: Text

Role Arn MediaConnect can assumes to create ENIs in customer's account

name :: Text

The name of the VPC Interface. This value must be unique within the current flow.

newVpcInterfaceRequest

Arguments

:: Text
$sel:subnetId:VpcInterfaceRequest'

-> Text
$sel:roleArn:VpcInterfaceRequest'

-> Text
$sel:name:VpcInterfaceRequest'

-> VpcInterfaceRequest

Create a value of VpcInterfaceRequest with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:networkInterfaceType:VpcInterfaceRequest', vpcInterfaceRequest_networkInterfaceType - The type of network interface. If this value is not included in the request, MediaConnect uses ENA as the networkInterfaceType.

$sel:subnetId:VpcInterfaceRequest', vpcInterfaceRequest_subnetId - Subnet must be in the AZ of the Flow

$sel:securityGroupIds:VpcInterfaceRequest', vpcInterfaceRequest_securityGroupIds - Security Group IDs to be used on ENI.

$sel:roleArn:VpcInterfaceRequest', vpcInterfaceRequest_roleArn - Role Arn MediaConnect can assumes to create ENIs in customer's account

$sel:name:VpcInterfaceRequest', vpcInterfaceRequest_name - The name of the VPC Interface. This value must be unique within the current flow.

vpcInterfaceRequest_networkInterfaceType :: Lens' VpcInterfaceRequest (Maybe NetworkInterfaceType)

The type of network interface. If this value is not included in the request, MediaConnect uses ENA as the networkInterfaceType.

vpcInterfaceRequest_subnetId :: Lens' VpcInterfaceRequest Text

Subnet must be in the AZ of the Flow

vpcInterfaceRequest_securityGroupIds :: Lens' VpcInterfaceRequest [Text]

Security Group IDs to be used on ENI.

vpcInterfaceRequest_roleArn :: Lens' VpcInterfaceRequest Text

Role Arn MediaConnect can assumes to create ENIs in customer's account

vpcInterfaceRequest_name :: Lens' VpcInterfaceRequest Text

The name of the VPC Interface. This value must be unique within the current flow.