Amazonka.MediaConnect.UpdateFlowEntitlement

Description

You can change an entitlement's description, subscribers, and encryption. If you change the subscribers, the service will remove the outputs that are are used by the subscribers that are removed.

Creating a Request

data UpdateFlowEntitlement

The entitlement fields that you want to update.

See: newUpdateFlowEntitlement smart constructor.

Constructors

UpdateFlowEntitlement'

Fields

encryption :: Maybe UpdateEncryption

The type of encryption that will be used on the output associated with this entitlement.

entitlementStatus :: Maybe EntitlementStatus

An indication of whether you want to enable the entitlement to allow access, or disable it to stop streaming content to the subscriber’s flow temporarily. If you don’t specify the entitlementStatus field in your request, MediaConnect leaves the value unchanged.

description :: Maybe Text

A description of the entitlement. This description appears only on the AWS Elemental MediaConnect console and will not be seen by the subscriber or end user.

subscribers :: Maybe [Text]

The AWS account IDs that you want to share your content with. The receiving accounts (subscribers) will be allowed to create their own flow using your content as the source.

flowArn :: Text

The flow that is associated with the entitlement that you want to update.

entitlementArn :: Text

The ARN of the entitlement that you want to update.

newUpdateFlowEntitlement

Arguments

:: Text
$sel:flowArn:UpdateFlowEntitlement'

-> Text
$sel:entitlementArn:UpdateFlowEntitlement'

-> UpdateFlowEntitlement

Create a value of UpdateFlowEntitlement with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:encryption:UpdateFlowEntitlement', updateFlowEntitlement_encryption - The type of encryption that will be used on the output associated with this entitlement.

$sel:entitlementStatus:UpdateFlowEntitlement', updateFlowEntitlement_entitlementStatus - An indication of whether you want to enable the entitlement to allow access, or disable it to stop streaming content to the subscriber’s flow temporarily. If you don’t specify the entitlementStatus field in your request, MediaConnect leaves the value unchanged.

$sel:description:UpdateFlowEntitlement', updateFlowEntitlement_description - A description of the entitlement. This description appears only on the AWS Elemental MediaConnect console and will not be seen by the subscriber or end user.

$sel:subscribers:UpdateFlowEntitlement', updateFlowEntitlement_subscribers - The AWS account IDs that you want to share your content with. The receiving accounts (subscribers) will be allowed to create their own flow using your content as the source.

$sel:flowArn:UpdateFlowEntitlement', updateFlowEntitlement_flowArn - The flow that is associated with the entitlement that you want to update.

$sel:entitlementArn:UpdateFlowEntitlement', updateFlowEntitlement_entitlementArn - The ARN of the entitlement that you want to update.

Request Lenses

updateFlowEntitlement_encryption :: Lens' UpdateFlowEntitlement (Maybe UpdateEncryption)

The type of encryption that will be used on the output associated with this entitlement.

updateFlowEntitlement_entitlementStatus :: Lens' UpdateFlowEntitlement (Maybe EntitlementStatus)

An indication of whether you want to enable the entitlement to allow access, or disable it to stop streaming content to the subscriber’s flow temporarily. If you don’t specify the entitlementStatus field in your request, MediaConnect leaves the value unchanged.

updateFlowEntitlement_description :: Lens' UpdateFlowEntitlement (Maybe Text)

A description of the entitlement. This description appears only on the AWS Elemental MediaConnect console and will not be seen by the subscriber or end user.

updateFlowEntitlement_subscribers :: Lens' UpdateFlowEntitlement (Maybe [Text])

The AWS account IDs that you want to share your content with. The receiving accounts (subscribers) will be allowed to create their own flow using your content as the source.

updateFlowEntitlement_flowArn :: Lens' UpdateFlowEntitlement Text

The flow that is associated with the entitlement that you want to update.

updateFlowEntitlement_entitlementArn :: Lens' UpdateFlowEntitlement Text

The ARN of the entitlement that you want to update.

Destructuring the Response

data UpdateFlowEntitlementResponse

See: newUpdateFlowEntitlementResponse smart constructor.

Constructors

UpdateFlowEntitlementResponse'

Fields

flowArn :: Maybe Text

The ARN of the flow that this entitlement was granted on.

entitlement :: Maybe Entitlement

The new configuration of the entitlement that you updated.

httpStatus :: Int

The response's http status code.

newUpdateFlowEntitlementResponse

Arguments

:: Int
$sel:httpStatus:UpdateFlowEntitlementResponse'

-> UpdateFlowEntitlementResponse

Create a value of UpdateFlowEntitlementResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:flowArn:UpdateFlowEntitlement', updateFlowEntitlementResponse_flowArn - The ARN of the flow that this entitlement was granted on.

$sel:entitlement:UpdateFlowEntitlementResponse', updateFlowEntitlementResponse_entitlement - The new configuration of the entitlement that you updated.

$sel:httpStatus:UpdateFlowEntitlementResponse', updateFlowEntitlementResponse_httpStatus - The response's http status code.

Response Lenses

updateFlowEntitlementResponse_flowArn :: Lens' UpdateFlowEntitlementResponse (Maybe Text)

The ARN of the flow that this entitlement was granted on.

updateFlowEntitlementResponse_entitlement :: Lens' UpdateFlowEntitlementResponse (Maybe Entitlement)

The new configuration of the entitlement that you updated.

updateFlowEntitlementResponse_httpStatus :: Lens' UpdateFlowEntitlementResponse Int

The response's http status code.