Amazonka.MediaConvert.CreateJob

Description

Create a new transcoding job. For information about jobs and job settings, see the User Guide at http://docs.aws.amazon.com/mediaconvert/latest/ug/what-is.html

Creating a Request

data CreateJob

See: newCreateJob smart constructor.

Constructors

CreateJob'

Fields

jobTemplate :: Maybe Text

Optional. When you create a job, you can either specify a job template or specify the transcoding settings individually.

accelerationSettings :: Maybe AccelerationSettings

Optional. Accelerated transcoding can significantly speed up jobs with long, visually complex content. Outputs that use this feature incur pro-tier pricing. For information about feature limitations, see the AWS Elemental MediaConvert User Guide.

priority :: Maybe Int

Optional. Specify the relative priority for this job. In any given queue, the service begins processing the job with the highest value first. When more than one job has the same priority, the service begins processing the job that you submitted first. If you don't specify a priority, the service uses the default value 0.

statusUpdateInterval :: Maybe StatusUpdateInterval

Optional. Specify how often MediaConvert sends STATUS_UPDATE events to Amazon CloudWatch Events. Set the interval, in seconds, between status updates. MediaConvert sends an update at this interval from the time the service begins processing your job to the time it completes the transcode or encounters an error.

hopDestinations :: Maybe [HopDestination]

Optional. Use queue hopping to avoid overly long waits in the backlog of the queue that you submit your job to. Specify an alternate queue and the maximum time that your job will wait in the initial queue before hopping. For more information about this feature, see the AWS Elemental MediaConvert User Guide.

simulateReservedQueue :: Maybe SimulateReservedQueue

Optional. Enable this setting when you run a test job to estimate how many reserved transcoding slots (RTS) you need. When this is enabled, MediaConvert runs your job from an on-demand queue with similar performance to what you will see with one RTS in a reserved queue. This setting is disabled by default.

queue :: Maybe Text

Optional. When you create a job, you can specify a queue to send it to. If you don't specify, the job will go to the default queue. For more about queues, see the User Guide topic at https://docs.aws.amazon.com/mediaconvert/latest/ug/what-is.html.

userMetadata :: Maybe (HashMap Text Text)

Optional. User-defined metadata that you want to associate with an MediaConvert job. You specify metadata in key/value pairs. Use only for existing integrations or workflows that rely on job metadata tags. Otherwise, we recommend that you use standard AWS tags.

billingTagsSource :: Maybe BillingTagsSource

Optional. Choose a tag type that AWS Billing and Cost Management will use to sort your AWS Elemental MediaConvert costs on any billing report that you set up. Any transcoding outputs that don't have an associated tag will appear in your billing report unsorted. If you don't choose a valid value for this field, your job outputs will appear on the billing report unsorted.

clientRequestToken :: Maybe Text

Optional. Idempotency token for CreateJob operation.

tags :: Maybe (HashMap Text Text)

Optional. The tags that you want to add to the resource. You can tag resources with a key-value pair or with only a key. Use standard AWS tags on your job for automatic integration with AWS services and for custom integrations and workflows.

role' :: Text

Required. The IAM role you use for creating this job. For details about permissions, see the User Guide topic at the User Guide at https://docs.aws.amazon.com/mediaconvert/latest/ug/iam-role.html.

settings :: JobSettings

JobSettings contains all the transcode settings for a job.

newCreateJob

Arguments

:: Text
$sel:role':CreateJob'

-> JobSettings
$sel:settings:CreateJob'

-> CreateJob

Create a value of CreateJob with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:jobTemplate:CreateJob', createJob_jobTemplate - Optional. When you create a job, you can either specify a job template or specify the transcoding settings individually.

$sel:accelerationSettings:CreateJob', createJob_accelerationSettings - Optional. Accelerated transcoding can significantly speed up jobs with long, visually complex content. Outputs that use this feature incur pro-tier pricing. For information about feature limitations, see the AWS Elemental MediaConvert User Guide.

$sel:priority:CreateJob', createJob_priority - Optional. Specify the relative priority for this job. In any given queue, the service begins processing the job with the highest value first. When more than one job has the same priority, the service begins processing the job that you submitted first. If you don't specify a priority, the service uses the default value 0.

$sel:statusUpdateInterval:CreateJob', createJob_statusUpdateInterval - Optional. Specify how often MediaConvert sends STATUS_UPDATE events to Amazon CloudWatch Events. Set the interval, in seconds, between status updates. MediaConvert sends an update at this interval from the time the service begins processing your job to the time it completes the transcode or encounters an error.

$sel:hopDestinations:CreateJob', createJob_hopDestinations - Optional. Use queue hopping to avoid overly long waits in the backlog of the queue that you submit your job to. Specify an alternate queue and the maximum time that your job will wait in the initial queue before hopping. For more information about this feature, see the AWS Elemental MediaConvert User Guide.

$sel:simulateReservedQueue:CreateJob', createJob_simulateReservedQueue - Optional. Enable this setting when you run a test job to estimate how many reserved transcoding slots (RTS) you need. When this is enabled, MediaConvert runs your job from an on-demand queue with similar performance to what you will see with one RTS in a reserved queue. This setting is disabled by default.

$sel:queue:CreateJob', createJob_queue - Optional. When you create a job, you can specify a queue to send it to. If you don't specify, the job will go to the default queue. For more about queues, see the User Guide topic at https://docs.aws.amazon.com/mediaconvert/latest/ug/what-is.html.

$sel:userMetadata:CreateJob', createJob_userMetadata - Optional. User-defined metadata that you want to associate with an MediaConvert job. You specify metadata in key/value pairs. Use only for existing integrations or workflows that rely on job metadata tags. Otherwise, we recommend that you use standard AWS tags.

$sel:billingTagsSource:CreateJob', createJob_billingTagsSource - Optional. Choose a tag type that AWS Billing and Cost Management will use to sort your AWS Elemental MediaConvert costs on any billing report that you set up. Any transcoding outputs that don't have an associated tag will appear in your billing report unsorted. If you don't choose a valid value for this field, your job outputs will appear on the billing report unsorted.

$sel:clientRequestToken:CreateJob', createJob_clientRequestToken - Optional. Idempotency token for CreateJob operation.

$sel:tags:CreateJob', createJob_tags - Optional. The tags that you want to add to the resource. You can tag resources with a key-value pair or with only a key. Use standard AWS tags on your job for automatic integration with AWS services and for custom integrations and workflows.

$sel:role':CreateJob', createJob_role - Required. The IAM role you use for creating this job. For details about permissions, see the User Guide topic at the User Guide at https://docs.aws.amazon.com/mediaconvert/latest/ug/iam-role.html.

$sel:settings:CreateJob', createJob_settings - JobSettings contains all the transcode settings for a job.

Request Lenses

createJob_jobTemplate :: Lens' CreateJob (Maybe Text)

Optional. When you create a job, you can either specify a job template or specify the transcoding settings individually.

createJob_accelerationSettings :: Lens' CreateJob (Maybe AccelerationSettings)

Optional. Accelerated transcoding can significantly speed up jobs with long, visually complex content. Outputs that use this feature incur pro-tier pricing. For information about feature limitations, see the AWS Elemental MediaConvert User Guide.

createJob_priority :: Lens' CreateJob (Maybe Int)

Optional. Specify the relative priority for this job. In any given queue, the service begins processing the job with the highest value first. When more than one job has the same priority, the service begins processing the job that you submitted first. If you don't specify a priority, the service uses the default value 0.

createJob_statusUpdateInterval :: Lens' CreateJob (Maybe StatusUpdateInterval)

Optional. Specify how often MediaConvert sends STATUS_UPDATE events to Amazon CloudWatch Events. Set the interval, in seconds, between status updates. MediaConvert sends an update at this interval from the time the service begins processing your job to the time it completes the transcode or encounters an error.

createJob_hopDestinations :: Lens' CreateJob (Maybe [HopDestination])

Optional. Use queue hopping to avoid overly long waits in the backlog of the queue that you submit your job to. Specify an alternate queue and the maximum time that your job will wait in the initial queue before hopping. For more information about this feature, see the AWS Elemental MediaConvert User Guide.

createJob_simulateReservedQueue :: Lens' CreateJob (Maybe SimulateReservedQueue)

Optional. Enable this setting when you run a test job to estimate how many reserved transcoding slots (RTS) you need. When this is enabled, MediaConvert runs your job from an on-demand queue with similar performance to what you will see with one RTS in a reserved queue. This setting is disabled by default.

createJob_queue :: Lens' CreateJob (Maybe Text)

Optional. When you create a job, you can specify a queue to send it to. If you don't specify, the job will go to the default queue. For more about queues, see the User Guide topic at https://docs.aws.amazon.com/mediaconvert/latest/ug/what-is.html.

createJob_userMetadata :: Lens' CreateJob (Maybe (HashMap Text Text))

Optional. User-defined metadata that you want to associate with an MediaConvert job. You specify metadata in key/value pairs. Use only for existing integrations or workflows that rely on job metadata tags. Otherwise, we recommend that you use standard AWS tags.

createJob_billingTagsSource :: Lens' CreateJob (Maybe BillingTagsSource)

Optional. Choose a tag type that AWS Billing and Cost Management will use to sort your AWS Elemental MediaConvert costs on any billing report that you set up. Any transcoding outputs that don't have an associated tag will appear in your billing report unsorted. If you don't choose a valid value for this field, your job outputs will appear on the billing report unsorted.

createJob_clientRequestToken :: Lens' CreateJob (Maybe Text)

Optional. Idempotency token for CreateJob operation.

createJob_tags :: Lens' CreateJob (Maybe (HashMap Text Text))

Optional. The tags that you want to add to the resource. You can tag resources with a key-value pair or with only a key. Use standard AWS tags on your job for automatic integration with AWS services and for custom integrations and workflows.

createJob_role :: Lens' CreateJob Text

Required. The IAM role you use for creating this job. For details about permissions, see the User Guide topic at the User Guide at https://docs.aws.amazon.com/mediaconvert/latest/ug/iam-role.html.

createJob_settings :: Lens' CreateJob JobSettings

JobSettings contains all the transcode settings for a job.

Destructuring the Response

data CreateJobResponse

See: newCreateJobResponse smart constructor.

Constructors

CreateJobResponse'

Fields

job :: Maybe Job

Each job converts an input file into an output file or files. For more information, see the User Guide at https://docs.aws.amazon.com/mediaconvert/latest/ug/what-is.html

httpStatus :: Int

The response's http status code.

newCreateJobResponse

Arguments

:: Int
$sel:httpStatus:CreateJobResponse'

-> CreateJobResponse

Create a value of CreateJobResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:job:CreateJobResponse', createJobResponse_job - Each job converts an input file into an output file or files. For more information, see the User Guide at https://docs.aws.amazon.com/mediaconvert/latest/ug/what-is.html

$sel:httpStatus:CreateJobResponse', createJobResponse_httpStatus - The response's http status code.

Response Lenses

createJobResponse_job :: Lens' CreateJobResponse (Maybe Job)

Each job converts an input file into an output file or files. For more information, see the User Guide at https://docs.aws.amazon.com/mediaconvert/latest/ug/what-is.html

createJobResponse_httpStatus :: Lens' CreateJobResponse Int

The response's http status code.