Amazonka.MediaConvert.CreateQueue

Description

Create a new transcoding queue. For information about queues, see Working With Queues in the User Guide at https://docs.aws.amazon.com/mediaconvert/latest/ug/working-with-queues.html

Creating a Request

data CreateQueue

See: newCreateQueue smart constructor.

Constructors

CreateQueue'

Fields

status :: Maybe QueueStatus

Initial state of the queue. If you create a paused queue, then jobs in that queue won't begin.

pricingPlan :: Maybe PricingPlan

Specifies whether the pricing plan for the queue is on-demand or reserved. For on-demand, you pay per minute, billed in increments of .01 minute. For reserved, you pay for the transcoding capacity of the entire queue, regardless of how much or how little you use it. Reserved pricing requires a 12-month commitment. When you use the API to create a queue, the default is on-demand.

description :: Maybe Text

Optional. A description of the queue that you are creating.

reservationPlanSettings :: Maybe ReservationPlanSettings

Details about the pricing plan for your reserved queue. Required for reserved queues and not applicable to on-demand queues.

tags :: Maybe (HashMap Text Text)

The tags that you want to add to the resource. You can tag resources with a key-value pair or with only a key.

name :: Text

The name of the queue that you are creating.

newCreateQueue

Arguments

:: Text
$sel:name:CreateQueue'

-> CreateQueue

Create a value of CreateQueue with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:status:CreateQueue', createQueue_status - Initial state of the queue. If you create a paused queue, then jobs in that queue won't begin.

$sel:pricingPlan:CreateQueue', createQueue_pricingPlan - Specifies whether the pricing plan for the queue is on-demand or reserved. For on-demand, you pay per minute, billed in increments of .01 minute. For reserved, you pay for the transcoding capacity of the entire queue, regardless of how much or how little you use it. Reserved pricing requires a 12-month commitment. When you use the API to create a queue, the default is on-demand.

$sel:description:CreateQueue', createQueue_description - Optional. A description of the queue that you are creating.

$sel:reservationPlanSettings:CreateQueue', createQueue_reservationPlanSettings - Details about the pricing plan for your reserved queue. Required for reserved queues and not applicable to on-demand queues.

$sel:tags:CreateQueue', createQueue_tags - The tags that you want to add to the resource. You can tag resources with a key-value pair or with only a key.

$sel:name:CreateQueue', createQueue_name - The name of the queue that you are creating.

Request Lenses

createQueue_status :: Lens' CreateQueue (Maybe QueueStatus)

Initial state of the queue. If you create a paused queue, then jobs in that queue won't begin.

createQueue_pricingPlan :: Lens' CreateQueue (Maybe PricingPlan)

Specifies whether the pricing plan for the queue is on-demand or reserved. For on-demand, you pay per minute, billed in increments of .01 minute. For reserved, you pay for the transcoding capacity of the entire queue, regardless of how much or how little you use it. Reserved pricing requires a 12-month commitment. When you use the API to create a queue, the default is on-demand.

createQueue_description :: Lens' CreateQueue (Maybe Text)

Optional. A description of the queue that you are creating.

createQueue_reservationPlanSettings :: Lens' CreateQueue (Maybe ReservationPlanSettings)

Details about the pricing plan for your reserved queue. Required for reserved queues and not applicable to on-demand queues.

createQueue_tags :: Lens' CreateQueue (Maybe (HashMap Text Text))

The tags that you want to add to the resource. You can tag resources with a key-value pair or with only a key.

createQueue_name :: Lens' CreateQueue Text

The name of the queue that you are creating.

Destructuring the Response

data CreateQueueResponse

See: newCreateQueueResponse smart constructor.

Constructors

CreateQueueResponse'

Fields

queue :: Maybe Queue

You can use queues to manage the resources that are available to your AWS account for running multiple transcoding jobs at the same time. If you don't specify a queue, the service sends all jobs through the default queue. For more information, see https://docs.aws.amazon.com/mediaconvert/latest/ug/working-with-queues.html.

httpStatus :: Int

The response's http status code.

newCreateQueueResponse

Arguments

:: Int
$sel:httpStatus:CreateQueueResponse'

-> CreateQueueResponse

Create a value of CreateQueueResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:queue:CreateQueueResponse', createQueueResponse_queue - You can use queues to manage the resources that are available to your AWS account for running multiple transcoding jobs at the same time. If you don't specify a queue, the service sends all jobs through the default queue. For more information, see https://docs.aws.amazon.com/mediaconvert/latest/ug/working-with-queues.html.

$sel:httpStatus:CreateQueueResponse', createQueueResponse_httpStatus - The response's http status code.

Response Lenses

createQueueResponse_queue :: Lens' CreateQueueResponse (Maybe Queue)

You can use queues to manage the resources that are available to your AWS account for running multiple transcoding jobs at the same time. If you don't specify a queue, the service sends all jobs through the default queue. For more information, see https://docs.aws.amazon.com/mediaconvert/latest/ug/working-with-queues.html.

createQueueResponse_httpStatus :: Lens' CreateQueueResponse Int

The response's http status code.