Amazonka.MediaConvert.ListPresets

Description

Retrieve a JSON array of up to twenty of your presets. This will return the presets themselves, not just a list of them. To retrieve the next twenty presets, use the nextToken string returned with the array.

This operation returns paginated results.

Creating a Request

data ListPresets

See: newListPresets smart constructor.

Constructors

ListPresets'

Fields

category :: Maybe Text

Optionally, specify a preset category to limit responses to only presets from that category.

listBy :: Maybe PresetListBy

Optional. When you request a list of presets, you can choose to list them alphabetically by NAME or chronologically by CREATION_DATE. If you don't specify, the service will list them by name.

nextToken :: Maybe Text

Use this string, provided with the response to a previous request, to request the next batch of presets.

order :: Maybe Order

Optional. When you request lists of resources, you can specify whether they are sorted in ASCENDING or DESCENDING order. Default varies by resource.

maxResults :: Maybe Natural

Optional. Number of presets, up to twenty, that will be returned at one time

newListPresets :: ListPresets

Create a value of ListPresets with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:category:ListPresets', listPresets_category - Optionally, specify a preset category to limit responses to only presets from that category.

$sel:listBy:ListPresets', listPresets_listBy - Optional. When you request a list of presets, you can choose to list them alphabetically by NAME or chronologically by CREATION_DATE. If you don't specify, the service will list them by name.

$sel:nextToken:ListPresets', listPresets_nextToken - Use this string, provided with the response to a previous request, to request the next batch of presets.

$sel:order:ListPresets', listPresets_order - Optional. When you request lists of resources, you can specify whether they are sorted in ASCENDING or DESCENDING order. Default varies by resource.

$sel:maxResults:ListPresets', listPresets_maxResults - Optional. Number of presets, up to twenty, that will be returned at one time

Request Lenses

listPresets_category :: Lens' ListPresets (Maybe Text)

Optionally, specify a preset category to limit responses to only presets from that category.

listPresets_listBy :: Lens' ListPresets (Maybe PresetListBy)

Optional. When you request a list of presets, you can choose to list them alphabetically by NAME or chronologically by CREATION_DATE. If you don't specify, the service will list them by name.

listPresets_nextToken :: Lens' ListPresets (Maybe Text)

Use this string, provided with the response to a previous request, to request the next batch of presets.

listPresets_order :: Lens' ListPresets (Maybe Order)

Optional. When you request lists of resources, you can specify whether they are sorted in ASCENDING or DESCENDING order. Default varies by resource.

listPresets_maxResults :: Lens' ListPresets (Maybe Natural)

Optional. Number of presets, up to twenty, that will be returned at one time

Destructuring the Response

data ListPresetsResponse

See: newListPresetsResponse smart constructor.

Constructors

ListPresetsResponse'

Fields

presets :: Maybe [Preset]

List of presets

nextToken :: Maybe Text

Use this string to request the next batch of presets.

httpStatus :: Int

The response's http status code.

newListPresetsResponse

Arguments

:: Int
$sel:httpStatus:ListPresetsResponse'

-> ListPresetsResponse

Create a value of ListPresetsResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:presets:ListPresetsResponse', listPresetsResponse_presets - List of presets

$sel:nextToken:ListPresets', listPresetsResponse_nextToken - Use this string to request the next batch of presets.

$sel:httpStatus:ListPresetsResponse', listPresetsResponse_httpStatus - The response's http status code.

Response Lenses

listPresetsResponse_presets :: Lens' ListPresetsResponse (Maybe [Preset])

List of presets

listPresetsResponse_nextToken :: Lens' ListPresetsResponse (Maybe Text)

Use this string to request the next batch of presets.

listPresetsResponse_httpStatus :: Lens' ListPresetsResponse Int

The response's http status code.