Amazonka.MediaConvert.Types.AacSettings

Description

Documentation

data AacSettings

Required when you set (Codec) under (AudioDescriptions)>(CodecSettings) to the value AAC. The service accepts one of two mutually exclusive groups of AAC settings--VBR and CBR. To select one of these modes, set the value of Bitrate control mode (rateControlMode) to "VBR" or "CBR". In VBR mode, you control the audio quality with the setting VBR quality (vbrQuality). In CBR mode, you use the setting Bitrate (bitrate). Defaults and valid values depend on the rate control mode.

See: newAacSettings smart constructor.

Constructors

AacSettings'

Fields

audioDescriptionBroadcasterMix :: Maybe AacAudioDescriptionBroadcasterMix

Choose BROADCASTER_MIXED_AD when the input contains pre-mixed main audio + audio description (AD) as a stereo pair. The value for AudioType will be set to 3, which signals to downstream systems that this stream contains "broadcaster mixed AD". Note that the input received by the encoder must contain pre-mixed audio; the encoder does not perform the mixing. When you choose BROADCASTER_MIXED_AD, the encoder ignores any values you provide in AudioType and FollowInputAudioType. Choose NORMAL when the input does not contain pre-mixed audio + audio description (AD). In this case, the encoder will use any values you provide for AudioType and FollowInputAudioType.

rawFormat :: Maybe AacRawFormat

Enables LATM/LOAS AAC output. Note that if you use LATM/LOAS AAC in an output, you must choose "No container" for the output container.

codingMode :: Maybe AacCodingMode

Mono (Audio Description), Mono, Stereo, or 5.1 channel layout. Valid values depend on rate control mode and profile. "1.0 - Audio Description (Receiver Mix)" setting receives a stereo description plus control track and emits a mono AAC encode of the description track, with control data emitted in the PES header as per ETSI TS 101 154 Annex E.

rateControlMode :: Maybe AacRateControlMode

Rate Control Mode.

sampleRate :: Maybe Natural

Sample rate in Hz. Valid values depend on rate control mode and profile.

specification :: Maybe AacSpecification

Use MPEG-2 AAC instead of MPEG-4 AAC audio for raw or MPEG-2 Transport Stream containers.

codecProfile :: Maybe AacCodecProfile

AAC Profile.

bitrate :: Maybe Natural

Specify the average bitrate in bits per second. The set of valid values for this setting is: 6000, 8000, 10000, 12000, 14000, 16000, 20000, 24000, 28000, 32000, 40000, 48000, 56000, 64000, 80000, 96000, 112000, 128000, 160000, 192000, 224000, 256000, 288000, 320000, 384000, 448000, 512000, 576000, 640000, 768000, 896000, 1024000. The value you set is also constrained by the values that you choose for Profile (codecProfile), Bitrate control mode (codingMode), and Sample rate (sampleRate). Default values depend on Bitrate control mode and Profile.

vbrQuality :: Maybe AacVbrQuality

VBR Quality Level - Only used if rate_control_mode is VBR.

newAacSettings :: AacSettings

Create a value of AacSettings with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:audioDescriptionBroadcasterMix:AacSettings', aacSettings_audioDescriptionBroadcasterMix - Choose BROADCASTER_MIXED_AD when the input contains pre-mixed main audio + audio description (AD) as a stereo pair. The value for AudioType will be set to 3, which signals to downstream systems that this stream contains "broadcaster mixed AD". Note that the input received by the encoder must contain pre-mixed audio; the encoder does not perform the mixing. When you choose BROADCASTER_MIXED_AD, the encoder ignores any values you provide in AudioType and FollowInputAudioType. Choose NORMAL when the input does not contain pre-mixed audio + audio description (AD). In this case, the encoder will use any values you provide for AudioType and FollowInputAudioType.

$sel:rawFormat:AacSettings', aacSettings_rawFormat - Enables LATM/LOAS AAC output. Note that if you use LATM/LOAS AAC in an output, you must choose "No container" for the output container.

$sel:codingMode:AacSettings', aacSettings_codingMode - Mono (Audio Description), Mono, Stereo, or 5.1 channel layout. Valid values depend on rate control mode and profile. "1.0 - Audio Description (Receiver Mix)" setting receives a stereo description plus control track and emits a mono AAC encode of the description track, with control data emitted in the PES header as per ETSI TS 101 154 Annex E.

$sel:rateControlMode:AacSettings', aacSettings_rateControlMode - Rate Control Mode.

$sel:sampleRate:AacSettings', aacSettings_sampleRate - Sample rate in Hz. Valid values depend on rate control mode and profile.

$sel:specification:AacSettings', aacSettings_specification - Use MPEG-2 AAC instead of MPEG-4 AAC audio for raw or MPEG-2 Transport Stream containers.

$sel:codecProfile:AacSettings', aacSettings_codecProfile - AAC Profile.

$sel:bitrate:AacSettings', aacSettings_bitrate - Specify the average bitrate in bits per second. The set of valid values for this setting is: 6000, 8000, 10000, 12000, 14000, 16000, 20000, 24000, 28000, 32000, 40000, 48000, 56000, 64000, 80000, 96000, 112000, 128000, 160000, 192000, 224000, 256000, 288000, 320000, 384000, 448000, 512000, 576000, 640000, 768000, 896000, 1024000. The value you set is also constrained by the values that you choose for Profile (codecProfile), Bitrate control mode (codingMode), and Sample rate (sampleRate). Default values depend on Bitrate control mode and Profile.

$sel:vbrQuality:AacSettings', aacSettings_vbrQuality - VBR Quality Level - Only used if rate_control_mode is VBR.

aacSettings_audioDescriptionBroadcasterMix :: Lens' AacSettings (Maybe AacAudioDescriptionBroadcasterMix)

Choose BROADCASTER_MIXED_AD when the input contains pre-mixed main audio + audio description (AD) as a stereo pair. The value for AudioType will be set to 3, which signals to downstream systems that this stream contains "broadcaster mixed AD". Note that the input received by the encoder must contain pre-mixed audio; the encoder does not perform the mixing. When you choose BROADCASTER_MIXED_AD, the encoder ignores any values you provide in AudioType and FollowInputAudioType. Choose NORMAL when the input does not contain pre-mixed audio + audio description (AD). In this case, the encoder will use any values you provide for AudioType and FollowInputAudioType.

aacSettings_rawFormat :: Lens' AacSettings (Maybe AacRawFormat)

Enables LATM/LOAS AAC output. Note that if you use LATM/LOAS AAC in an output, you must choose "No container" for the output container.

aacSettings_codingMode :: Lens' AacSettings (Maybe AacCodingMode)

Mono (Audio Description), Mono, Stereo, or 5.1 channel layout. Valid values depend on rate control mode and profile. "1.0 - Audio Description (Receiver Mix)" setting receives a stereo description plus control track and emits a mono AAC encode of the description track, with control data emitted in the PES header as per ETSI TS 101 154 Annex E.

aacSettings_rateControlMode :: Lens' AacSettings (Maybe AacRateControlMode)

Rate Control Mode.

aacSettings_sampleRate :: Lens' AacSettings (Maybe Natural)

Sample rate in Hz. Valid values depend on rate control mode and profile.

aacSettings_specification :: Lens' AacSettings (Maybe AacSpecification)

Use MPEG-2 AAC instead of MPEG-4 AAC audio for raw or MPEG-2 Transport Stream containers.

aacSettings_codecProfile :: Lens' AacSettings (Maybe AacCodecProfile)

AAC Profile.

aacSettings_bitrate :: Lens' AacSettings (Maybe Natural)

Specify the average bitrate in bits per second. The set of valid values for this setting is: 6000, 8000, 10000, 12000, 14000, 16000, 20000, 24000, 28000, 32000, 40000, 48000, 56000, 64000, 80000, 96000, 112000, 128000, 160000, 192000, 224000, 256000, 288000, 320000, 384000, 448000, 512000, 576000, 640000, 768000, 896000, 1024000. The value you set is also constrained by the values that you choose for Profile (codecProfile), Bitrate control mode (codingMode), and Sample rate (sampleRate). Default values depend on Bitrate control mode and Profile.

aacSettings_vbrQuality :: Lens' AacSettings (Maybe AacVbrQuality)

VBR Quality Level - Only used if rate_control_mode is VBR.