Amazonka.MediaConvert.Types.AudioNormalizationSettings

Description

Documentation

data AudioNormalizationSettings

Advanced audio normalization settings. Ignore these settings unless you need to comply with a loudness standard.

See: newAudioNormalizationSettings smart constructor.

Constructors

AudioNormalizationSettings'

Fields

algorithmControl :: Maybe AudioNormalizationAlgorithmControl

When enabled the output audio is corrected using the chosen algorithm. If disabled, the audio will be measured but not adjusted.

targetLkfs :: Maybe Double

When you use Audio normalization (AudioNormalizationSettings), optionally use this setting to specify a target loudness. If you don't specify a value here, the encoder chooses a value for you, based on the algorithm that you choose for Algorithm (algorithm). If you choose algorithm 1770-1, the encoder will choose -24 LKFS; otherwise, the encoder will choose -23 LKFS.

peakCalculation :: Maybe AudioNormalizationPeakCalculation

If set to TRUE_PEAK, calculate and log the TruePeak for each output's audio track loudness.

correctionGateLevel :: Maybe Int

Content measuring above this level will be corrected to the target level. Content measuring below this level will not be corrected.

algorithm :: Maybe AudioNormalizationAlgorithm

Choose one of the following audio normalization algorithms: ITU-R BS.1770-1: Ungated loudness. A measurement of ungated average loudness for an entire piece of content, suitable for measurement of short-form content under ATSC recommendation A/85. Supports up to 5.1 audio channels. ITU-R BS.1770-2: Gated loudness. A measurement of gated average loudness compliant with the requirements of EBU-R128. Supports up to 5.1 audio channels. ITU-R BS.1770-3: Modified peak. The same loudness measurement algorithm as 1770-2, with an updated true peak measurement. ITU-R BS.1770-4: Higher channel count. Allows for more audio channels than the other algorithms, including configurations such as 7.1.

loudnessLogging :: Maybe AudioNormalizationLoudnessLogging

If set to LOG, log each output's audio track loudness to a CSV file.

newAudioNormalizationSettings :: AudioNormalizationSettings

Create a value of AudioNormalizationSettings with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:algorithmControl:AudioNormalizationSettings', audioNormalizationSettings_algorithmControl - When enabled the output audio is corrected using the chosen algorithm. If disabled, the audio will be measured but not adjusted.

$sel:targetLkfs:AudioNormalizationSettings', audioNormalizationSettings_targetLkfs - When you use Audio normalization (AudioNormalizationSettings), optionally use this setting to specify a target loudness. If you don't specify a value here, the encoder chooses a value for you, based on the algorithm that you choose for Algorithm (algorithm). If you choose algorithm 1770-1, the encoder will choose -24 LKFS; otherwise, the encoder will choose -23 LKFS.

$sel:peakCalculation:AudioNormalizationSettings', audioNormalizationSettings_peakCalculation - If set to TRUE_PEAK, calculate and log the TruePeak for each output's audio track loudness.

$sel:correctionGateLevel:AudioNormalizationSettings', audioNormalizationSettings_correctionGateLevel - Content measuring above this level will be corrected to the target level. Content measuring below this level will not be corrected.

$sel:algorithm:AudioNormalizationSettings', audioNormalizationSettings_algorithm - Choose one of the following audio normalization algorithms: ITU-R BS.1770-1: Ungated loudness. A measurement of ungated average loudness for an entire piece of content, suitable for measurement of short-form content under ATSC recommendation A/85. Supports up to 5.1 audio channels. ITU-R BS.1770-2: Gated loudness. A measurement of gated average loudness compliant with the requirements of EBU-R128. Supports up to 5.1 audio channels. ITU-R BS.1770-3: Modified peak. The same loudness measurement algorithm as 1770-2, with an updated true peak measurement. ITU-R BS.1770-4: Higher channel count. Allows for more audio channels than the other algorithms, including configurations such as 7.1.

$sel:loudnessLogging:AudioNormalizationSettings', audioNormalizationSettings_loudnessLogging - If set to LOG, log each output's audio track loudness to a CSV file.

audioNormalizationSettings_algorithmControl :: Lens' AudioNormalizationSettings (Maybe AudioNormalizationAlgorithmControl)

When enabled the output audio is corrected using the chosen algorithm. If disabled, the audio will be measured but not adjusted.

audioNormalizationSettings_targetLkfs :: Lens' AudioNormalizationSettings (Maybe Double)

When you use Audio normalization (AudioNormalizationSettings), optionally use this setting to specify a target loudness. If you don't specify a value here, the encoder chooses a value for you, based on the algorithm that you choose for Algorithm (algorithm). If you choose algorithm 1770-1, the encoder will choose -24 LKFS; otherwise, the encoder will choose -23 LKFS.

audioNormalizationSettings_peakCalculation :: Lens' AudioNormalizationSettings (Maybe AudioNormalizationPeakCalculation)

If set to TRUE_PEAK, calculate and log the TruePeak for each output's audio track loudness.

audioNormalizationSettings_correctionGateLevel :: Lens' AudioNormalizationSettings (Maybe Int)

Content measuring above this level will be corrected to the target level. Content measuring below this level will not be corrected.

audioNormalizationSettings_algorithm :: Lens' AudioNormalizationSettings (Maybe AudioNormalizationAlgorithm)

Choose one of the following audio normalization algorithms: ITU-R BS.1770-1: Ungated loudness. A measurement of ungated average loudness for an entire piece of content, suitable for measurement of short-form content under ATSC recommendation A/85. Supports up to 5.1 audio channels. ITU-R BS.1770-2: Gated loudness. A measurement of gated average loudness compliant with the requirements of EBU-R128. Supports up to 5.1 audio channels. ITU-R BS.1770-3: Modified peak. The same loudness measurement algorithm as 1770-2, with an updated true peak measurement. ITU-R BS.1770-4: Higher channel count. Allows for more audio channels than the other algorithms, including configurations such as 7.1.

audioNormalizationSettings_loudnessLogging :: Lens' AudioNormalizationSettings (Maybe AudioNormalizationLoudnessLogging)

If set to LOG, log each output's audio track loudness to a CSV file.