Amazonka.MediaConvert.Types.CaptionSourceSettings

Description

Documentation

data CaptionSourceSettings

If your input captions are SCC, TTML, STL, SMI, SRT, or IMSC in an xml file, specify the URI of the input captions source file. If your input captions are IMSC in an IMF package, use TrackSourceSettings instead of FileSoureSettings.

See: newCaptionSourceSettings smart constructor.

Constructors

CaptionSourceSettings'

Fields

teletextSourceSettings :: Maybe TeletextSourceSettings

Settings specific to Teletext caption sources, including Page number.

sourceType :: Maybe CaptionSourceType

Use Source (SourceType) to identify the format of your input captions. The service cannot auto-detect caption format.

fileSourceSettings :: Maybe FileSourceSettings

If your input captions are SCC, SMI, SRT, STL, TTML, WebVTT, or IMSC 1.1 in an xml file, specify the URI of the input caption source file. If your caption source is IMSC in an IMF package, use TrackSourceSettings instead of FileSoureSettings.

webvttHlsSourceSettings :: Maybe WebvttHlsSourceSettings

Settings specific to WebVTT sources in HLS alternative rendition group. Specify the properties (renditionGroupId, renditionName or renditionLanguageCode) to identify the unique subtitle track among the alternative rendition groups present in the HLS manifest. If no unique track is found, or multiple tracks match the specified properties, the job fails. If there is only one subtitle track in the rendition group, the settings can be left empty and the default subtitle track will be chosen. If your caption source is a sidecar file, use FileSourceSettings instead of WebvttHlsSourceSettings.

dvbSubSourceSettings :: Maybe DvbSubSourceSettings

DVB Sub Source Settings

trackSourceSettings :: Maybe TrackSourceSettings

Settings specific to caption sources that are specified by track number. Currently, this is only IMSC captions in an IMF package. If your caption source is IMSC 1.1 in a separate xml file, use FileSourceSettings instead of TrackSourceSettings.

ancillarySourceSettings :: Maybe AncillarySourceSettings

Settings for ancillary captions source.

embeddedSourceSettings :: Maybe EmbeddedSourceSettings

Settings for embedded captions Source

newCaptionSourceSettings :: CaptionSourceSettings

Create a value of CaptionSourceSettings with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:teletextSourceSettings:CaptionSourceSettings', captionSourceSettings_teletextSourceSettings - Settings specific to Teletext caption sources, including Page number.

$sel:sourceType:CaptionSourceSettings', captionSourceSettings_sourceType - Use Source (SourceType) to identify the format of your input captions. The service cannot auto-detect caption format.

$sel:fileSourceSettings:CaptionSourceSettings', captionSourceSettings_fileSourceSettings - If your input captions are SCC, SMI, SRT, STL, TTML, WebVTT, or IMSC 1.1 in an xml file, specify the URI of the input caption source file. If your caption source is IMSC in an IMF package, use TrackSourceSettings instead of FileSoureSettings.

$sel:webvttHlsSourceSettings:CaptionSourceSettings', captionSourceSettings_webvttHlsSourceSettings - Settings specific to WebVTT sources in HLS alternative rendition group. Specify the properties (renditionGroupId, renditionName or renditionLanguageCode) to identify the unique subtitle track among the alternative rendition groups present in the HLS manifest. If no unique track is found, or multiple tracks match the specified properties, the job fails. If there is only one subtitle track in the rendition group, the settings can be left empty and the default subtitle track will be chosen. If your caption source is a sidecar file, use FileSourceSettings instead of WebvttHlsSourceSettings.

$sel:dvbSubSourceSettings:CaptionSourceSettings', captionSourceSettings_dvbSubSourceSettings - DVB Sub Source Settings

$sel:trackSourceSettings:CaptionSourceSettings', captionSourceSettings_trackSourceSettings - Settings specific to caption sources that are specified by track number. Currently, this is only IMSC captions in an IMF package. If your caption source is IMSC 1.1 in a separate xml file, use FileSourceSettings instead of TrackSourceSettings.

$sel:ancillarySourceSettings:CaptionSourceSettings', captionSourceSettings_ancillarySourceSettings - Settings for ancillary captions source.

$sel:embeddedSourceSettings:CaptionSourceSettings', captionSourceSettings_embeddedSourceSettings - Settings for embedded captions Source

captionSourceSettings_teletextSourceSettings :: Lens' CaptionSourceSettings (Maybe TeletextSourceSettings)

Settings specific to Teletext caption sources, including Page number.

captionSourceSettings_sourceType :: Lens' CaptionSourceSettings (Maybe CaptionSourceType)

Use Source (SourceType) to identify the format of your input captions. The service cannot auto-detect caption format.

captionSourceSettings_fileSourceSettings :: Lens' CaptionSourceSettings (Maybe FileSourceSettings)

If your input captions are SCC, SMI, SRT, STL, TTML, WebVTT, or IMSC 1.1 in an xml file, specify the URI of the input caption source file. If your caption source is IMSC in an IMF package, use TrackSourceSettings instead of FileSoureSettings.

captionSourceSettings_webvttHlsSourceSettings :: Lens' CaptionSourceSettings (Maybe WebvttHlsSourceSettings)

Settings specific to WebVTT sources in HLS alternative rendition group. Specify the properties (renditionGroupId, renditionName or renditionLanguageCode) to identify the unique subtitle track among the alternative rendition groups present in the HLS manifest. If no unique track is found, or multiple tracks match the specified properties, the job fails. If there is only one subtitle track in the rendition group, the settings can be left empty and the default subtitle track will be chosen. If your caption source is a sidecar file, use FileSourceSettings instead of WebvttHlsSourceSettings.

captionSourceSettings_dvbSubSourceSettings :: Lens' CaptionSourceSettings (Maybe DvbSubSourceSettings)

DVB Sub Source Settings

captionSourceSettings_trackSourceSettings :: Lens' CaptionSourceSettings (Maybe TrackSourceSettings)

Settings specific to caption sources that are specified by track number. Currently, this is only IMSC captions in an IMF package. If your caption source is IMSC 1.1 in a separate xml file, use FileSourceSettings instead of TrackSourceSettings.

captionSourceSettings_ancillarySourceSettings :: Lens' CaptionSourceSettings (Maybe AncillarySourceSettings)

Settings for ancillary captions source.

captionSourceSettings_embeddedSourceSettings :: Lens' CaptionSourceSettings (Maybe EmbeddedSourceSettings)

Settings for embedded captions Source