Amazonka.MediaConvert.Types.CmafEncryptionSettings

Description

Documentation

data CmafEncryptionSettings

Settings for CMAF encryption

See: newCmafEncryptionSettings smart constructor.

Constructors

CmafEncryptionSettings'

Fields

encryptionMethod :: Maybe CmafEncryptionType

Specify the encryption scheme that you want the service to use when encrypting your CMAF segments. Choose AES-CBC subsample (SAMPLE-AES) or AES_CTR (AES-CTR).

constantInitializationVector :: Maybe Text

This is a 128-bit, 16-byte hex value represented by a 32-character text string. If this parameter is not set then the Initialization Vector will follow the segment number by default.

type' :: Maybe CmafKeyProviderType

Specify whether your DRM encryption key is static or from a key provider that follows the SPEKE standard. For more information about SPEKE, see https://docs.aws.amazon.com/speke/latest/documentation/what-is-speke.html.

staticKeyProvider :: Maybe StaticKeyProvider

Use these settings to set up encryption with a static key provider.

spekeKeyProvider :: Maybe SpekeKeyProviderCmaf

If your output group type is CMAF, use these settings when doing DRM encryption with a SPEKE-compliant key provider. If your output group type is HLS, DASH, or Microsoft Smooth, use the SpekeKeyProvider settings instead.

initializationVectorInManifest :: Maybe CmafInitializationVectorInManifest

When you use DRM with CMAF outputs, choose whether the service writes the 128-bit encryption initialization vector in the HLS and DASH manifests.

newCmafEncryptionSettings :: CmafEncryptionSettings

Create a value of CmafEncryptionSettings with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:encryptionMethod:CmafEncryptionSettings', cmafEncryptionSettings_encryptionMethod - Specify the encryption scheme that you want the service to use when encrypting your CMAF segments. Choose AES-CBC subsample (SAMPLE-AES) or AES_CTR (AES-CTR).

$sel:constantInitializationVector:CmafEncryptionSettings', cmafEncryptionSettings_constantInitializationVector - This is a 128-bit, 16-byte hex value represented by a 32-character text string. If this parameter is not set then the Initialization Vector will follow the segment number by default.

$sel:type':CmafEncryptionSettings', cmafEncryptionSettings_type - Specify whether your DRM encryption key is static or from a key provider that follows the SPEKE standard. For more information about SPEKE, see https://docs.aws.amazon.com/speke/latest/documentation/what-is-speke.html.

$sel:staticKeyProvider:CmafEncryptionSettings', cmafEncryptionSettings_staticKeyProvider - Use these settings to set up encryption with a static key provider.

$sel:spekeKeyProvider:CmafEncryptionSettings', cmafEncryptionSettings_spekeKeyProvider - If your output group type is CMAF, use these settings when doing DRM encryption with a SPEKE-compliant key provider. If your output group type is HLS, DASH, or Microsoft Smooth, use the SpekeKeyProvider settings instead.

$sel:initializationVectorInManifest:CmafEncryptionSettings', cmafEncryptionSettings_initializationVectorInManifest - When you use DRM with CMAF outputs, choose whether the service writes the 128-bit encryption initialization vector in the HLS and DASH manifests.

cmafEncryptionSettings_encryptionMethod :: Lens' CmafEncryptionSettings (Maybe CmafEncryptionType)

Specify the encryption scheme that you want the service to use when encrypting your CMAF segments. Choose AES-CBC subsample (SAMPLE-AES) or AES_CTR (AES-CTR).

cmafEncryptionSettings_constantInitializationVector :: Lens' CmafEncryptionSettings (Maybe Text)

This is a 128-bit, 16-byte hex value represented by a 32-character text string. If this parameter is not set then the Initialization Vector will follow the segment number by default.

cmafEncryptionSettings_type :: Lens' CmafEncryptionSettings (Maybe CmafKeyProviderType)

Specify whether your DRM encryption key is static or from a key provider that follows the SPEKE standard. For more information about SPEKE, see https://docs.aws.amazon.com/speke/latest/documentation/what-is-speke.html.

cmafEncryptionSettings_staticKeyProvider :: Lens' CmafEncryptionSettings (Maybe StaticKeyProvider)

Use these settings to set up encryption with a static key provider.

cmafEncryptionSettings_spekeKeyProvider :: Lens' CmafEncryptionSettings (Maybe SpekeKeyProviderCmaf)

If your output group type is CMAF, use these settings when doing DRM encryption with a SPEKE-compliant key provider. If your output group type is HLS, DASH, or Microsoft Smooth, use the SpekeKeyProvider settings instead.

cmafEncryptionSettings_initializationVectorInManifest :: Lens' CmafEncryptionSettings (Maybe CmafInitializationVectorInManifest)

When you use DRM with CMAF outputs, choose whether the service writes the 128-bit encryption initialization vector in the HLS and DASH manifests.