Amazonka.MediaConvert.Types.DashIsoImageBasedTrickPlaySettings

Description

Documentation

data DashIsoImageBasedTrickPlaySettings

Tile and thumbnail settings applicable when imageBasedTrickPlay is ADVANCED

See: newDashIsoImageBasedTrickPlaySettings smart constructor.

Constructors

DashIsoImageBasedTrickPlaySettings'

Fields

tileWidth :: Maybe Natural

Number of thumbnails in each row of a tile image. Set a value between 1 and 512.

thumbnailHeight :: Maybe Natural

Height of each thumbnail within each tile image, in pixels. Leave blank to maintain aspect ratio with thumbnail width. If following the aspect ratio would lead to a total tile height greater than 4096, then the job will be rejected. Must be divisible by 2.

intervalCadence :: Maybe DashIsoIntervalCadence

The cadence MediaConvert follows for generating thumbnails. If set to FOLLOW_IFRAME, MediaConvert generates thumbnails for each IDR frame in the output (matching the GOP cadence). If set to FOLLOW_CUSTOM, MediaConvert generates thumbnails according to the interval you specify in thumbnailInterval.

thumbnailWidth :: Maybe Natural

Width of each thumbnail within each tile image, in pixels. Default is 312. Must be divisible by 8.

thumbnailInterval :: Maybe Double

Enter the interval, in seconds, that MediaConvert uses to generate thumbnails. If the interval you enter doesn't align with the output frame rate, MediaConvert automatically rounds the interval to align with the output frame rate. For example, if the output frame rate is 29.97 frames per second and you enter 5, MediaConvert uses a 150 frame interval to generate thumbnails.

tileHeight :: Maybe Natural

Number of thumbnails in each column of a tile image. Set a value between 2 and 2048. Must be divisible by 2.

newDashIsoImageBasedTrickPlaySettings :: DashIsoImageBasedTrickPlaySettings

Create a value of DashIsoImageBasedTrickPlaySettings with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:tileWidth:DashIsoImageBasedTrickPlaySettings', dashIsoImageBasedTrickPlaySettings_tileWidth - Number of thumbnails in each row of a tile image. Set a value between 1 and 512.

$sel:thumbnailHeight:DashIsoImageBasedTrickPlaySettings', dashIsoImageBasedTrickPlaySettings_thumbnailHeight - Height of each thumbnail within each tile image, in pixels. Leave blank to maintain aspect ratio with thumbnail width. If following the aspect ratio would lead to a total tile height greater than 4096, then the job will be rejected. Must be divisible by 2.

$sel:intervalCadence:DashIsoImageBasedTrickPlaySettings', dashIsoImageBasedTrickPlaySettings_intervalCadence - The cadence MediaConvert follows for generating thumbnails. If set to FOLLOW_IFRAME, MediaConvert generates thumbnails for each IDR frame in the output (matching the GOP cadence). If set to FOLLOW_CUSTOM, MediaConvert generates thumbnails according to the interval you specify in thumbnailInterval.

$sel:thumbnailWidth:DashIsoImageBasedTrickPlaySettings', dashIsoImageBasedTrickPlaySettings_thumbnailWidth - Width of each thumbnail within each tile image, in pixels. Default is 312. Must be divisible by 8.

$sel:thumbnailInterval:DashIsoImageBasedTrickPlaySettings', dashIsoImageBasedTrickPlaySettings_thumbnailInterval - Enter the interval, in seconds, that MediaConvert uses to generate thumbnails. If the interval you enter doesn't align with the output frame rate, MediaConvert automatically rounds the interval to align with the output frame rate. For example, if the output frame rate is 29.97 frames per second and you enter 5, MediaConvert uses a 150 frame interval to generate thumbnails.

$sel:tileHeight:DashIsoImageBasedTrickPlaySettings', dashIsoImageBasedTrickPlaySettings_tileHeight - Number of thumbnails in each column of a tile image. Set a value between 2 and 2048. Must be divisible by 2.

dashIsoImageBasedTrickPlaySettings_tileWidth :: Lens' DashIsoImageBasedTrickPlaySettings (Maybe Natural)

Number of thumbnails in each row of a tile image. Set a value between 1 and 512.

dashIsoImageBasedTrickPlaySettings_thumbnailHeight :: Lens' DashIsoImageBasedTrickPlaySettings (Maybe Natural)

Height of each thumbnail within each tile image, in pixels. Leave blank to maintain aspect ratio with thumbnail width. If following the aspect ratio would lead to a total tile height greater than 4096, then the job will be rejected. Must be divisible by 2.

dashIsoImageBasedTrickPlaySettings_intervalCadence :: Lens' DashIsoImageBasedTrickPlaySettings (Maybe DashIsoIntervalCadence)

The cadence MediaConvert follows for generating thumbnails. If set to FOLLOW_IFRAME, MediaConvert generates thumbnails for each IDR frame in the output (matching the GOP cadence). If set to FOLLOW_CUSTOM, MediaConvert generates thumbnails according to the interval you specify in thumbnailInterval.

dashIsoImageBasedTrickPlaySettings_thumbnailWidth :: Lens' DashIsoImageBasedTrickPlaySettings (Maybe Natural)

Width of each thumbnail within each tile image, in pixels. Default is 312. Must be divisible by 8.

dashIsoImageBasedTrickPlaySettings_thumbnailInterval :: Lens' DashIsoImageBasedTrickPlaySettings (Maybe Double)

Enter the interval, in seconds, that MediaConvert uses to generate thumbnails. If the interval you enter doesn't align with the output frame rate, MediaConvert automatically rounds the interval to align with the output frame rate. For example, if the output frame rate is 29.97 frames per second and you enter 5, MediaConvert uses a 150 frame interval to generate thumbnails.

dashIsoImageBasedTrickPlaySettings_tileHeight :: Lens' DashIsoImageBasedTrickPlaySettings (Maybe Natural)

Number of thumbnails in each column of a tile image. Set a value between 2 and 2048. Must be divisible by 2.