Amazonka.MediaConvert.Types.M2tsSettings

Description

Documentation

data M2tsSettings

MPEG-2 TS container settings. These apply to outputs in a File output group when the output's container (ContainerType) is MPEG-2 Transport Stream (M2TS). In these assets, data is organized by the program map table (PMT). Each transport stream program contains subsets of data, including audio, video, and metadata. Each of these subsets of data has a numerical label called a packet identifier (PID). Each transport stream program corresponds to one MediaConvert output. The PMT lists the types of data in a program along with their PID. Downstream systems and players use the program map table to look up the PID for each type of data it accesses and then uses the PIDs to locate specific data within the asset.

See: newM2tsSettings smart constructor.

Constructors

M2tsSettings'

Fields

pmtPid :: Maybe Natural

Specify the packet identifier (PID) for the program map table (PMT) itself. Default is 480.

videoPid :: Maybe Natural

Specify the packet identifier (PID) of the elementary video stream in the transport stream.

bufferModel :: Maybe M2tsBufferModel

Controls what buffer model to use for accurate interleaving. If set to MULTIPLEX, use multiplex buffer model. If set to NONE, this can lead to lower latency, but low-memory devices may not be able to play back the stream without interruptions.

programNumber :: Maybe Natural

Use Program number (programNumber) to specify the program number used in the program map table (PMT) for this output. Default is 1. Program numbers and program map tables are parts of MPEG-2 transport stream containers, used for organizing data.

scte35Pid :: Maybe Natural

Specify the packet identifier (PID) of the SCTE-35 stream in the transport stream.

minEbpInterval :: Maybe Natural

When set, enforces that Encoder Boundary Points do not come within the specified time interval of each other by looking ahead at input video. If another EBP is going to come in within the specified time interval, the current EBP is not emitted, and the segment is "stretched" to the next marker. The lookahead value does not add latency to the system. The Live Event must be configured elsewhere to create sufficient latency to make the lookahead accurate.

transportStreamId :: Maybe Natural

Specify the ID for the transport stream itself in the program map table for this output. Transport stream IDs and program map tables are parts of MPEG-2 transport stream containers, used for organizing data.

maxPcrInterval :: Maybe Natural

Specify the maximum time, in milliseconds, between Program Clock References (PCRs) inserted into the transport stream.

fragmentTime :: Maybe Double

The length, in seconds, of each fragment. Only used with EBP markers.

privateMetadataPid :: Maybe Natural

Specify the packet identifier (PID) of the private metadata stream. Default is 503.

scte35Esam :: Maybe M2tsScte35Esam

Include this in your job settings to put SCTE-35 markers in your HLS and transport stream outputs at the insertion points that you specify in an ESAM XML document. Provide the document in the setting SCC XML (sccXml).

audioDuration :: Maybe M2tsAudioDuration

Specify this setting only when your output will be consumed by a downstream repackaging workflow that is sensitive to very small duration differences between video and audio. For this situation, choose Match video duration (MATCH_VIDEO_DURATION). In all other cases, keep the default value, Default codec duration (DEFAULT_CODEC_DURATION). When you choose Match video duration, MediaConvert pads the output audio streams with silence or trims them to ensure that the total duration of each audio stream is at least as long as the total duration of the video stream. After padding or trimming, the audio stream duration is no more than one frame longer than the video stream. MediaConvert applies audio padding or trimming only to the end of the last segment of the output. For unsegmented outputs, MediaConvert adds padding only to the end of the file. When you keep the default value, any minor discrepancies between audio and video duration will depend on your output audio codec.

pmtInterval :: Maybe Natural

Specify the number of milliseconds between instances of the program map table (PMT) in the output transport stream.

dvbSdtSettings :: Maybe DvbSdtSettings

Use these settings to insert a DVB Service Description Table (SDT) in the transport stream of this output. When you work directly in your JSON job specification, include this object only when your job has a transport stream output and the container settings contain the object M2tsSettings.

nullPacketBitrate :: Maybe Double

Value in bits per second of extra null packets to insert into the transport stream. This can be used if a downstream encryption system requires periodic null packets.

audioBufferModel :: Maybe M2tsAudioBufferModel

Selects between the DVB and ATSC buffer models for Dolby Digital audio.

timedMetadataPid :: Maybe Natural

Specify the packet identifier (PID) for timed metadata in this output. Default is 502.

audioFramesPerPes :: Maybe Natural

The number of audio frames to insert for each PES packet.

pcrPid :: Maybe Natural

Specify the packet identifier (PID) for the program clock reference (PCR) in this output. If you do not specify a value, the service will use the value for Video PID (VideoPid).

segmentationMarkers :: Maybe M2tsSegmentationMarkers

Inserts segmentation markers at each segmentation_time period. rai_segstart sets the Random Access Indicator bit in the adaptation field. rai_adapt sets the RAI bit and adds the current timecode in the private data bytes. psi_segstart inserts PAT and PMT tables at the start of segments. ebp adds Encoder Boundary Point information to the adaptation field as per OpenCable specification OC-SP-EBP-I01-130118. ebp_legacy adds Encoder Boundary Point information to the adaptation field using a legacy proprietary format.

dvbSubPids :: Maybe [Natural]

Specify the packet identifiers (PIDs) for DVB subtitle data included in this output. Specify multiple PIDs as a JSON array. Default is the range 460-479.

scte35Source :: Maybe M2tsScte35Source

For SCTE-35 markers from your input-- Choose Passthrough (PASSTHROUGH) if you want SCTE-35 markers that appear in your input to also appear in this output. Choose None (NONE) if you don't want SCTE-35 markers in this output. For SCTE-35 markers from an ESAM XML document-- Choose None (NONE). Also provide the ESAM XML as a string in the setting Signal processing notification XML (sccXml). Also enable ESAM SCTE-35 (include the property scte35Esam).

patInterval :: Maybe Natural

The number of milliseconds between instances of this table in the output transport stream.

forceTsVideoEbpOrder :: Maybe M2tsForceTsVideoEbpOrder

Keep the default value (DEFAULT) unless you know that your audio EBP markers are incorrectly appearing before your video EBP markers. To correct this problem, set this value to Force (FORCE).

esRateInPes :: Maybe M2tsEsRateInPes

Controls whether to include the ES Rate field in the PES header.

bitrate :: Maybe Natural

Specify the output bitrate of the transport stream in bits per second. Setting to 0 lets the muxer automatically determine the appropriate bitrate. Other common values are 3750000, 7500000, and 15000000.

audioPids :: Maybe [Natural]

Specify the packet identifiers (PIDs) for any elementary audio streams you include in this output. Specify multiple PIDs as a JSON array. Default is the range 482-492.

dvbTeletextPid :: Maybe Natural

Specify the packet identifier (PID) for DVB teletext data you include in this output. Default is 499.

nielsenId3 :: Maybe M2tsNielsenId3

If INSERT, Nielsen inaudible tones for media tracking will be detected in the input audio and an equivalent ID3 tag will be inserted in the output.

dataPTSControl :: Maybe M2tsDataPtsControl

If you select ALIGN_TO_VIDEO, MediaConvert writes captions and data packets with Presentation Timestamp (PTS) values greater than or equal to the first video packet PTS (MediaConvert drops captions and data packets with lesser PTS values). Keep the default value (AUTO) to allow all PTS values.

segmentationTime :: Maybe Double

Specify the length, in seconds, of each segment. Required unless markers is set to _none_.

ebpAudioInterval :: Maybe M2tsEbpAudioInterval

When set to VIDEO_AND_FIXED_INTERVALS, audio EBP markers will be added to partitions 3 and 4. The interval between these additional markers will be fixed, and will be slightly shorter than the video EBP marker interval. When set to VIDEO_INTERVAL, these additional markers will not be inserted. Only applicable when EBP segmentation markers are is selected (segmentationMarkers is EBP or EBP_LEGACY).

dvbNitSettings :: Maybe DvbNitSettings

Use these settings to insert a DVB Network Information Table (NIT) in the transport stream of this output. When you work directly in your JSON job specification, include this object only when your job has a transport stream output and the container settings contain the object M2tsSettings.

pcrControl :: Maybe M2tsPcrControl

When set to PCR_EVERY_PES_PACKET, a Program Clock Reference value is inserted for every Packetized Elementary Stream (PES) header. This is effective only when the PCR PID is the same as the video or audio elementary stream.

ebpPlacement :: Maybe M2tsEbpPlacement

Selects which PIDs to place EBP markers on. They can either be placed only on the video PID, or on both the video PID and all audio PIDs. Only applicable when EBP segmentation markers are is selected (segmentationMarkers is EBP or EBP_LEGACY).

rateMode :: Maybe M2tsRateMode

When set to CBR, inserts null packets into transport stream to fill specified bitrate. When set to VBR, the bitrate setting acts as the maximum bitrate, but the output will not be padded up to that bitrate.

segmentationStyle :: Maybe M2tsSegmentationStyle

The segmentation style parameter controls how segmentation markers are inserted into the transport stream. With avails, it is possible that segments may be truncated, which can influence where future segmentation markers are inserted. When a segmentation style of "reset_cadence" is selected and a segment is truncated due to an avail, we will reset the segmentation cadence. This means the subsequent segment will have a duration of of $segmentation_time seconds. When a segmentation style of "maintain_cadence" is selected and a segment is truncated due to an avail, we will not reset the segmentation cadence. This means the subsequent segment will likely be truncated as well. However, all segments after that will have a duration of $segmentation_time seconds. Note that EBP lookahead is a slight exception to this rule.

dvbTdtSettings :: Maybe DvbTdtSettings

Use these settings to insert a DVB Time and Date Table (TDT) in the transport stream of this output. When you work directly in your JSON job specification, include this object only when your job has a transport stream output and the container settings contain the object M2tsSettings.

newM2tsSettings :: M2tsSettings

Create a value of M2tsSettings with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:pmtPid:M2tsSettings', m2tsSettings_pmtPid - Specify the packet identifier (PID) for the program map table (PMT) itself. Default is 480.

$sel:videoPid:M2tsSettings', m2tsSettings_videoPid - Specify the packet identifier (PID) of the elementary video stream in the transport stream.

$sel:bufferModel:M2tsSettings', m2tsSettings_bufferModel - Controls what buffer model to use for accurate interleaving. If set to MULTIPLEX, use multiplex buffer model. If set to NONE, this can lead to lower latency, but low-memory devices may not be able to play back the stream without interruptions.

$sel:programNumber:M2tsSettings', m2tsSettings_programNumber - Use Program number (programNumber) to specify the program number used in the program map table (PMT) for this output. Default is 1. Program numbers and program map tables are parts of MPEG-2 transport stream containers, used for organizing data.

$sel:scte35Pid:M2tsSettings', m2tsSettings_scte35Pid - Specify the packet identifier (PID) of the SCTE-35 stream in the transport stream.

$sel:minEbpInterval:M2tsSettings', m2tsSettings_minEbpInterval - When set, enforces that Encoder Boundary Points do not come within the specified time interval of each other by looking ahead at input video. If another EBP is going to come in within the specified time interval, the current EBP is not emitted, and the segment is "stretched" to the next marker. The lookahead value does not add latency to the system. The Live Event must be configured elsewhere to create sufficient latency to make the lookahead accurate.

$sel:transportStreamId:M2tsSettings', m2tsSettings_transportStreamId - Specify the ID for the transport stream itself in the program map table for this output. Transport stream IDs and program map tables are parts of MPEG-2 transport stream containers, used for organizing data.

$sel:maxPcrInterval:M2tsSettings', m2tsSettings_maxPcrInterval - Specify the maximum time, in milliseconds, between Program Clock References (PCRs) inserted into the transport stream.

$sel:fragmentTime:M2tsSettings', m2tsSettings_fragmentTime - The length, in seconds, of each fragment. Only used with EBP markers.

$sel:privateMetadataPid:M2tsSettings', m2tsSettings_privateMetadataPid - Specify the packet identifier (PID) of the private metadata stream. Default is 503.

$sel:scte35Esam:M2tsSettings', m2tsSettings_scte35Esam - Include this in your job settings to put SCTE-35 markers in your HLS and transport stream outputs at the insertion points that you specify in an ESAM XML document. Provide the document in the setting SCC XML (sccXml).

$sel:audioDuration:M2tsSettings', m2tsSettings_audioDuration - Specify this setting only when your output will be consumed by a downstream repackaging workflow that is sensitive to very small duration differences between video and audio. For this situation, choose Match video duration (MATCH_VIDEO_DURATION). In all other cases, keep the default value, Default codec duration (DEFAULT_CODEC_DURATION). When you choose Match video duration, MediaConvert pads the output audio streams with silence or trims them to ensure that the total duration of each audio stream is at least as long as the total duration of the video stream. After padding or trimming, the audio stream duration is no more than one frame longer than the video stream. MediaConvert applies audio padding or trimming only to the end of the last segment of the output. For unsegmented outputs, MediaConvert adds padding only to the end of the file. When you keep the default value, any minor discrepancies between audio and video duration will depend on your output audio codec.

$sel:pmtInterval:M2tsSettings', m2tsSettings_pmtInterval - Specify the number of milliseconds between instances of the program map table (PMT) in the output transport stream.

$sel:dvbSdtSettings:M2tsSettings', m2tsSettings_dvbSdtSettings - Use these settings to insert a DVB Service Description Table (SDT) in the transport stream of this output. When you work directly in your JSON job specification, include this object only when your job has a transport stream output and the container settings contain the object M2tsSettings.

$sel:nullPacketBitrate:M2tsSettings', m2tsSettings_nullPacketBitrate - Value in bits per second of extra null packets to insert into the transport stream. This can be used if a downstream encryption system requires periodic null packets.

$sel:audioBufferModel:M2tsSettings', m2tsSettings_audioBufferModel - Selects between the DVB and ATSC buffer models for Dolby Digital audio.

$sel:timedMetadataPid:M2tsSettings', m2tsSettings_timedMetadataPid - Specify the packet identifier (PID) for timed metadata in this output. Default is 502.

$sel:audioFramesPerPes:M2tsSettings', m2tsSettings_audioFramesPerPes - The number of audio frames to insert for each PES packet.

$sel:pcrPid:M2tsSettings', m2tsSettings_pcrPid - Specify the packet identifier (PID) for the program clock reference (PCR) in this output. If you do not specify a value, the service will use the value for Video PID (VideoPid).

$sel:segmentationMarkers:M2tsSettings', m2tsSettings_segmentationMarkers - Inserts segmentation markers at each segmentation_time period. rai_segstart sets the Random Access Indicator bit in the adaptation field. rai_adapt sets the RAI bit and adds the current timecode in the private data bytes. psi_segstart inserts PAT and PMT tables at the start of segments. ebp adds Encoder Boundary Point information to the adaptation field as per OpenCable specification OC-SP-EBP-I01-130118. ebp_legacy adds Encoder Boundary Point information to the adaptation field using a legacy proprietary format.

$sel:dvbSubPids:M2tsSettings', m2tsSettings_dvbSubPids - Specify the packet identifiers (PIDs) for DVB subtitle data included in this output. Specify multiple PIDs as a JSON array. Default is the range 460-479.

$sel:scte35Source:M2tsSettings', m2tsSettings_scte35Source - For SCTE-35 markers from your input-- Choose Passthrough (PASSTHROUGH) if you want SCTE-35 markers that appear in your input to also appear in this output. Choose None (NONE) if you don't want SCTE-35 markers in this output. For SCTE-35 markers from an ESAM XML document-- Choose None (NONE). Also provide the ESAM XML as a string in the setting Signal processing notification XML (sccXml). Also enable ESAM SCTE-35 (include the property scte35Esam).

$sel:patInterval:M2tsSettings', m2tsSettings_patInterval - The number of milliseconds between instances of this table in the output transport stream.

$sel:forceTsVideoEbpOrder:M2tsSettings', m2tsSettings_forceTsVideoEbpOrder - Keep the default value (DEFAULT) unless you know that your audio EBP markers are incorrectly appearing before your video EBP markers. To correct this problem, set this value to Force (FORCE).

$sel:esRateInPes:M2tsSettings', m2tsSettings_esRateInPes - Controls whether to include the ES Rate field in the PES header.

$sel:bitrate:M2tsSettings', m2tsSettings_bitrate - Specify the output bitrate of the transport stream in bits per second. Setting to 0 lets the muxer automatically determine the appropriate bitrate. Other common values are 3750000, 7500000, and 15000000.

$sel:audioPids:M2tsSettings', m2tsSettings_audioPids - Specify the packet identifiers (PIDs) for any elementary audio streams you include in this output. Specify multiple PIDs as a JSON array. Default is the range 482-492.

$sel:dvbTeletextPid:M2tsSettings', m2tsSettings_dvbTeletextPid - Specify the packet identifier (PID) for DVB teletext data you include in this output. Default is 499.

$sel:nielsenId3:M2tsSettings', m2tsSettings_nielsenId3 - If INSERT, Nielsen inaudible tones for media tracking will be detected in the input audio and an equivalent ID3 tag will be inserted in the output.

$sel:dataPTSControl:M2tsSettings', m2tsSettings_dataPTSControl - If you select ALIGN_TO_VIDEO, MediaConvert writes captions and data packets with Presentation Timestamp (PTS) values greater than or equal to the first video packet PTS (MediaConvert drops captions and data packets with lesser PTS values). Keep the default value (AUTO) to allow all PTS values.

$sel:segmentationTime:M2tsSettings', m2tsSettings_segmentationTime - Specify the length, in seconds, of each segment. Required unless markers is set to _none_.

$sel:ebpAudioInterval:M2tsSettings', m2tsSettings_ebpAudioInterval - When set to VIDEO_AND_FIXED_INTERVALS, audio EBP markers will be added to partitions 3 and 4. The interval between these additional markers will be fixed, and will be slightly shorter than the video EBP marker interval. When set to VIDEO_INTERVAL, these additional markers will not be inserted. Only applicable when EBP segmentation markers are is selected (segmentationMarkers is EBP or EBP_LEGACY).

$sel:dvbNitSettings:M2tsSettings', m2tsSettings_dvbNitSettings - Use these settings to insert a DVB Network Information Table (NIT) in the transport stream of this output. When you work directly in your JSON job specification, include this object only when your job has a transport stream output and the container settings contain the object M2tsSettings.

$sel:pcrControl:M2tsSettings', m2tsSettings_pcrControl - When set to PCR_EVERY_PES_PACKET, a Program Clock Reference value is inserted for every Packetized Elementary Stream (PES) header. This is effective only when the PCR PID is the same as the video or audio elementary stream.

$sel:ebpPlacement:M2tsSettings', m2tsSettings_ebpPlacement - Selects which PIDs to place EBP markers on. They can either be placed only on the video PID, or on both the video PID and all audio PIDs. Only applicable when EBP segmentation markers are is selected (segmentationMarkers is EBP or EBP_LEGACY).

$sel:rateMode:M2tsSettings', m2tsSettings_rateMode - When set to CBR, inserts null packets into transport stream to fill specified bitrate. When set to VBR, the bitrate setting acts as the maximum bitrate, but the output will not be padded up to that bitrate.

$sel:segmentationStyle:M2tsSettings', m2tsSettings_segmentationStyle - The segmentation style parameter controls how segmentation markers are inserted into the transport stream. With avails, it is possible that segments may be truncated, which can influence where future segmentation markers are inserted. When a segmentation style of "reset_cadence" is selected and a segment is truncated due to an avail, we will reset the segmentation cadence. This means the subsequent segment will have a duration of of $segmentation_time seconds. When a segmentation style of "maintain_cadence" is selected and a segment is truncated due to an avail, we will not reset the segmentation cadence. This means the subsequent segment will likely be truncated as well. However, all segments after that will have a duration of $segmentation_time seconds. Note that EBP lookahead is a slight exception to this rule.

$sel:dvbTdtSettings:M2tsSettings', m2tsSettings_dvbTdtSettings - Use these settings to insert a DVB Time and Date Table (TDT) in the transport stream of this output. When you work directly in your JSON job specification, include this object only when your job has a transport stream output and the container settings contain the object M2tsSettings.

m2tsSettings_pmtPid :: Lens' M2tsSettings (Maybe Natural)

Specify the packet identifier (PID) for the program map table (PMT) itself. Default is 480.

m2tsSettings_videoPid :: Lens' M2tsSettings (Maybe Natural)

Specify the packet identifier (PID) of the elementary video stream in the transport stream.

m2tsSettings_bufferModel :: Lens' M2tsSettings (Maybe M2tsBufferModel)

Controls what buffer model to use for accurate interleaving. If set to MULTIPLEX, use multiplex buffer model. If set to NONE, this can lead to lower latency, but low-memory devices may not be able to play back the stream without interruptions.

m2tsSettings_programNumber :: Lens' M2tsSettings (Maybe Natural)

Use Program number (programNumber) to specify the program number used in the program map table (PMT) for this output. Default is 1. Program numbers and program map tables are parts of MPEG-2 transport stream containers, used for organizing data.

m2tsSettings_scte35Pid :: Lens' M2tsSettings (Maybe Natural)

Specify the packet identifier (PID) of the SCTE-35 stream in the transport stream.

m2tsSettings_minEbpInterval :: Lens' M2tsSettings (Maybe Natural)

When set, enforces that Encoder Boundary Points do not come within the specified time interval of each other by looking ahead at input video. If another EBP is going to come in within the specified time interval, the current EBP is not emitted, and the segment is "stretched" to the next marker. The lookahead value does not add latency to the system. The Live Event must be configured elsewhere to create sufficient latency to make the lookahead accurate.

m2tsSettings_transportStreamId :: Lens' M2tsSettings (Maybe Natural)

Specify the ID for the transport stream itself in the program map table for this output. Transport stream IDs and program map tables are parts of MPEG-2 transport stream containers, used for organizing data.

m2tsSettings_maxPcrInterval :: Lens' M2tsSettings (Maybe Natural)

Specify the maximum time, in milliseconds, between Program Clock References (PCRs) inserted into the transport stream.

m2tsSettings_fragmentTime :: Lens' M2tsSettings (Maybe Double)

The length, in seconds, of each fragment. Only used with EBP markers.

m2tsSettings_privateMetadataPid :: Lens' M2tsSettings (Maybe Natural)

Specify the packet identifier (PID) of the private metadata stream. Default is 503.

m2tsSettings_scte35Esam :: Lens' M2tsSettings (Maybe M2tsScte35Esam)

Include this in your job settings to put SCTE-35 markers in your HLS and transport stream outputs at the insertion points that you specify in an ESAM XML document. Provide the document in the setting SCC XML (sccXml).

m2tsSettings_audioDuration :: Lens' M2tsSettings (Maybe M2tsAudioDuration)

Specify this setting only when your output will be consumed by a downstream repackaging workflow that is sensitive to very small duration differences between video and audio. For this situation, choose Match video duration (MATCH_VIDEO_DURATION). In all other cases, keep the default value, Default codec duration (DEFAULT_CODEC_DURATION). When you choose Match video duration, MediaConvert pads the output audio streams with silence or trims them to ensure that the total duration of each audio stream is at least as long as the total duration of the video stream. After padding or trimming, the audio stream duration is no more than one frame longer than the video stream. MediaConvert applies audio padding or trimming only to the end of the last segment of the output. For unsegmented outputs, MediaConvert adds padding only to the end of the file. When you keep the default value, any minor discrepancies between audio and video duration will depend on your output audio codec.

m2tsSettings_pmtInterval :: Lens' M2tsSettings (Maybe Natural)

Specify the number of milliseconds between instances of the program map table (PMT) in the output transport stream.

m2tsSettings_dvbSdtSettings :: Lens' M2tsSettings (Maybe DvbSdtSettings)

Use these settings to insert a DVB Service Description Table (SDT) in the transport stream of this output. When you work directly in your JSON job specification, include this object only when your job has a transport stream output and the container settings contain the object M2tsSettings.

m2tsSettings_nullPacketBitrate :: Lens' M2tsSettings (Maybe Double)

Value in bits per second of extra null packets to insert into the transport stream. This can be used if a downstream encryption system requires periodic null packets.

m2tsSettings_audioBufferModel :: Lens' M2tsSettings (Maybe M2tsAudioBufferModel)

Selects between the DVB and ATSC buffer models for Dolby Digital audio.

m2tsSettings_timedMetadataPid :: Lens' M2tsSettings (Maybe Natural)

Specify the packet identifier (PID) for timed metadata in this output. Default is 502.

m2tsSettings_audioFramesPerPes :: Lens' M2tsSettings (Maybe Natural)

The number of audio frames to insert for each PES packet.

m2tsSettings_pcrPid :: Lens' M2tsSettings (Maybe Natural)

Specify the packet identifier (PID) for the program clock reference (PCR) in this output. If you do not specify a value, the service will use the value for Video PID (VideoPid).

m2tsSettings_segmentationMarkers :: Lens' M2tsSettings (Maybe M2tsSegmentationMarkers)

Inserts segmentation markers at each segmentation_time period. rai_segstart sets the Random Access Indicator bit in the adaptation field. rai_adapt sets the RAI bit and adds the current timecode in the private data bytes. psi_segstart inserts PAT and PMT tables at the start of segments. ebp adds Encoder Boundary Point information to the adaptation field as per OpenCable specification OC-SP-EBP-I01-130118. ebp_legacy adds Encoder Boundary Point information to the adaptation field using a legacy proprietary format.

m2tsSettings_dvbSubPids :: Lens' M2tsSettings (Maybe [Natural])

Specify the packet identifiers (PIDs) for DVB subtitle data included in this output. Specify multiple PIDs as a JSON array. Default is the range 460-479.

m2tsSettings_scte35Source :: Lens' M2tsSettings (Maybe M2tsScte35Source)

For SCTE-35 markers from your input-- Choose Passthrough (PASSTHROUGH) if you want SCTE-35 markers that appear in your input to also appear in this output. Choose None (NONE) if you don't want SCTE-35 markers in this output. For SCTE-35 markers from an ESAM XML document-- Choose None (NONE). Also provide the ESAM XML as a string in the setting Signal processing notification XML (sccXml). Also enable ESAM SCTE-35 (include the property scte35Esam).

m2tsSettings_patInterval :: Lens' M2tsSettings (Maybe Natural)

The number of milliseconds between instances of this table in the output transport stream.

m2tsSettings_forceTsVideoEbpOrder :: Lens' M2tsSettings (Maybe M2tsForceTsVideoEbpOrder)

Keep the default value (DEFAULT) unless you know that your audio EBP markers are incorrectly appearing before your video EBP markers. To correct this problem, set this value to Force (FORCE).

m2tsSettings_esRateInPes :: Lens' M2tsSettings (Maybe M2tsEsRateInPes)

Controls whether to include the ES Rate field in the PES header.

m2tsSettings_bitrate :: Lens' M2tsSettings (Maybe Natural)

Specify the output bitrate of the transport stream in bits per second. Setting to 0 lets the muxer automatically determine the appropriate bitrate. Other common values are 3750000, 7500000, and 15000000.

m2tsSettings_audioPids :: Lens' M2tsSettings (Maybe [Natural])

Specify the packet identifiers (PIDs) for any elementary audio streams you include in this output. Specify multiple PIDs as a JSON array. Default is the range 482-492.

m2tsSettings_dvbTeletextPid :: Lens' M2tsSettings (Maybe Natural)

Specify the packet identifier (PID) for DVB teletext data you include in this output. Default is 499.

m2tsSettings_nielsenId3 :: Lens' M2tsSettings (Maybe M2tsNielsenId3)

If INSERT, Nielsen inaudible tones for media tracking will be detected in the input audio and an equivalent ID3 tag will be inserted in the output.

m2tsSettings_dataPTSControl :: Lens' M2tsSettings (Maybe M2tsDataPtsControl)

If you select ALIGN_TO_VIDEO, MediaConvert writes captions and data packets with Presentation Timestamp (PTS) values greater than or equal to the first video packet PTS (MediaConvert drops captions and data packets with lesser PTS values). Keep the default value (AUTO) to allow all PTS values.

m2tsSettings_segmentationTime :: Lens' M2tsSettings (Maybe Double)

Specify the length, in seconds, of each segment. Required unless markers is set to _none_.

m2tsSettings_ebpAudioInterval :: Lens' M2tsSettings (Maybe M2tsEbpAudioInterval)

When set to VIDEO_AND_FIXED_INTERVALS, audio EBP markers will be added to partitions 3 and 4. The interval between these additional markers will be fixed, and will be slightly shorter than the video EBP marker interval. When set to VIDEO_INTERVAL, these additional markers will not be inserted. Only applicable when EBP segmentation markers are is selected (segmentationMarkers is EBP or EBP_LEGACY).

m2tsSettings_dvbNitSettings :: Lens' M2tsSettings (Maybe DvbNitSettings)

Use these settings to insert a DVB Network Information Table (NIT) in the transport stream of this output. When you work directly in your JSON job specification, include this object only when your job has a transport stream output and the container settings contain the object M2tsSettings.

m2tsSettings_pcrControl :: Lens' M2tsSettings (Maybe M2tsPcrControl)

When set to PCR_EVERY_PES_PACKET, a Program Clock Reference value is inserted for every Packetized Elementary Stream (PES) header. This is effective only when the PCR PID is the same as the video or audio elementary stream.

m2tsSettings_ebpPlacement :: Lens' M2tsSettings (Maybe M2tsEbpPlacement)

Selects which PIDs to place EBP markers on. They can either be placed only on the video PID, or on both the video PID and all audio PIDs. Only applicable when EBP segmentation markers are is selected (segmentationMarkers is EBP or EBP_LEGACY).

m2tsSettings_rateMode :: Lens' M2tsSettings (Maybe M2tsRateMode)

When set to CBR, inserts null packets into transport stream to fill specified bitrate. When set to VBR, the bitrate setting acts as the maximum bitrate, but the output will not be padded up to that bitrate.

m2tsSettings_segmentationStyle :: Lens' M2tsSettings (Maybe M2tsSegmentationStyle)

The segmentation style parameter controls how segmentation markers are inserted into the transport stream. With avails, it is possible that segments may be truncated, which can influence where future segmentation markers are inserted. When a segmentation style of "reset_cadence" is selected and a segment is truncated due to an avail, we will reset the segmentation cadence. This means the subsequent segment will have a duration of of $segmentation_time seconds. When a segmentation style of "maintain_cadence" is selected and a segment is truncated due to an avail, we will not reset the segmentation cadence. This means the subsequent segment will likely be truncated as well. However, all segments after that will have a duration of $segmentation_time seconds. Note that EBP lookahead is a slight exception to this rule.

m2tsSettings_dvbTdtSettings :: Lens' M2tsSettings (Maybe DvbTdtSettings)

Use these settings to insert a DVB Time and Date Table (TDT) in the transport stream of this output. When you work directly in your JSON job specification, include this object only when your job has a transport stream output and the container settings contain the object M2tsSettings.