Amazonka.MediaConvert.Types.M3u8Settings

Description

Documentation

data M3u8Settings

These settings relate to the MPEG-2 transport stream (MPEG2-TS) container for the MPEG2-TS segments in your HLS outputs.

See: newM3u8Settings smart constructor.

Constructors

M3u8Settings'

Fields

pmtPid :: Maybe Natural

Packet Identifier (PID) for the Program Map Table (PMT) in the transport stream.

videoPid :: Maybe Natural

Packet Identifier (PID) of the elementary video stream in the transport stream.

programNumber :: Maybe Natural

The value of the program number field in the Program Map Table.

scte35Pid :: Maybe Natural

Packet Identifier (PID) of the SCTE-35 stream in the transport stream.

transportStreamId :: Maybe Natural

The value of the transport stream ID field in the Program Map Table.

maxPcrInterval :: Maybe Natural

Specify the maximum time, in milliseconds, between Program Clock References (PCRs) inserted into the transport stream.

privateMetadataPid :: Maybe Natural

Packet Identifier (PID) of the private metadata stream in the transport stream.

audioDuration :: Maybe M3u8AudioDuration

Specify this setting only when your output will be consumed by a downstream repackaging workflow that is sensitive to very small duration differences between video and audio. For this situation, choose Match video duration (MATCH_VIDEO_DURATION). In all other cases, keep the default value, Default codec duration (DEFAULT_CODEC_DURATION). When you choose Match video duration, MediaConvert pads the output audio streams with silence or trims them to ensure that the total duration of each audio stream is at least as long as the total duration of the video stream. After padding or trimming, the audio stream duration is no more than one frame longer than the video stream. MediaConvert applies audio padding or trimming only to the end of the last segment of the output. For unsegmented outputs, MediaConvert adds padding only to the end of the file. When you keep the default value, any minor discrepancies between audio and video duration will depend on your output audio codec.

pmtInterval :: Maybe Natural

The number of milliseconds between instances of this table in the output transport stream.

timedMetadataPid :: Maybe Natural

Packet Identifier (PID) of the timed metadata stream in the transport stream.

audioFramesPerPes :: Maybe Natural

The number of audio frames to insert for each PES packet.

pcrPid :: Maybe Natural

Packet Identifier (PID) of the Program Clock Reference (PCR) in the transport stream. When no value is given, the encoder will assign the same value as the Video PID.

timedMetadata :: Maybe TimedMetadata

Applies only to HLS outputs. Use this setting to specify whether the service inserts the ID3 timed metadata from the input in this output.

scte35Source :: Maybe M3u8Scte35Source

For SCTE-35 markers from your input-- Choose Passthrough (PASSTHROUGH) if you want SCTE-35 markers that appear in your input to also appear in this output. Choose None (NONE) if you don't want SCTE-35 markers in this output. For SCTE-35 markers from an ESAM XML document-- Choose None (NONE) if you don't want manifest conditioning. Choose Passthrough (PASSTHROUGH) and choose Ad markers (adMarkers) if you do want manifest conditioning. In both cases, also provide the ESAM XML as a string in the setting Signal processing notification XML (sccXml).

patInterval :: Maybe Natural

The number of milliseconds between instances of this table in the output transport stream.

audioPids :: Maybe [Natural]

Packet Identifier (PID) of the elementary audio stream(s) in the transport stream. Multiple values are accepted, and can be entered in ranges and/or by comma separation.

nielsenId3 :: Maybe M3u8NielsenId3

If INSERT, Nielsen inaudible tones for media tracking will be detected in the input audio and an equivalent ID3 tag will be inserted in the output.

dataPTSControl :: Maybe M3u8DataPtsControl

If you select ALIGN_TO_VIDEO, MediaConvert writes captions and data packets with Presentation Timestamp (PTS) values greater than or equal to the first video packet PTS (MediaConvert drops captions and data packets with lesser PTS values). Keep the default value (AUTO) to allow all PTS values.

pcrControl :: Maybe M3u8PcrControl

When set to PCR_EVERY_PES_PACKET a Program Clock Reference value is inserted for every Packetized Elementary Stream (PES) header. This parameter is effective only when the PCR PID is the same as the video or audio elementary stream.

newM3u8Settings :: M3u8Settings

Create a value of M3u8Settings with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:pmtPid:M3u8Settings', m3u8Settings_pmtPid - Packet Identifier (PID) for the Program Map Table (PMT) in the transport stream.

$sel:videoPid:M3u8Settings', m3u8Settings_videoPid - Packet Identifier (PID) of the elementary video stream in the transport stream.

$sel:programNumber:M3u8Settings', m3u8Settings_programNumber - The value of the program number field in the Program Map Table.

$sel:scte35Pid:M3u8Settings', m3u8Settings_scte35Pid - Packet Identifier (PID) of the SCTE-35 stream in the transport stream.

$sel:transportStreamId:M3u8Settings', m3u8Settings_transportStreamId - The value of the transport stream ID field in the Program Map Table.

$sel:maxPcrInterval:M3u8Settings', m3u8Settings_maxPcrInterval - Specify the maximum time, in milliseconds, between Program Clock References (PCRs) inserted into the transport stream.

$sel:privateMetadataPid:M3u8Settings', m3u8Settings_privateMetadataPid - Packet Identifier (PID) of the private metadata stream in the transport stream.

$sel:audioDuration:M3u8Settings', m3u8Settings_audioDuration - Specify this setting only when your output will be consumed by a downstream repackaging workflow that is sensitive to very small duration differences between video and audio. For this situation, choose Match video duration (MATCH_VIDEO_DURATION). In all other cases, keep the default value, Default codec duration (DEFAULT_CODEC_DURATION). When you choose Match video duration, MediaConvert pads the output audio streams with silence or trims them to ensure that the total duration of each audio stream is at least as long as the total duration of the video stream. After padding or trimming, the audio stream duration is no more than one frame longer than the video stream. MediaConvert applies audio padding or trimming only to the end of the last segment of the output. For unsegmented outputs, MediaConvert adds padding only to the end of the file. When you keep the default value, any minor discrepancies between audio and video duration will depend on your output audio codec.

$sel:pmtInterval:M3u8Settings', m3u8Settings_pmtInterval - The number of milliseconds between instances of this table in the output transport stream.

$sel:timedMetadataPid:M3u8Settings', m3u8Settings_timedMetadataPid - Packet Identifier (PID) of the timed metadata stream in the transport stream.

$sel:audioFramesPerPes:M3u8Settings', m3u8Settings_audioFramesPerPes - The number of audio frames to insert for each PES packet.

$sel:pcrPid:M3u8Settings', m3u8Settings_pcrPid - Packet Identifier (PID) of the Program Clock Reference (PCR) in the transport stream. When no value is given, the encoder will assign the same value as the Video PID.

$sel:timedMetadata:M3u8Settings', m3u8Settings_timedMetadata - Applies only to HLS outputs. Use this setting to specify whether the service inserts the ID3 timed metadata from the input in this output.

$sel:scte35Source:M3u8Settings', m3u8Settings_scte35Source - For SCTE-35 markers from your input-- Choose Passthrough (PASSTHROUGH) if you want SCTE-35 markers that appear in your input to also appear in this output. Choose None (NONE) if you don't want SCTE-35 markers in this output. For SCTE-35 markers from an ESAM XML document-- Choose None (NONE) if you don't want manifest conditioning. Choose Passthrough (PASSTHROUGH) and choose Ad markers (adMarkers) if you do want manifest conditioning. In both cases, also provide the ESAM XML as a string in the setting Signal processing notification XML (sccXml).

$sel:patInterval:M3u8Settings', m3u8Settings_patInterval - The number of milliseconds between instances of this table in the output transport stream.

$sel:audioPids:M3u8Settings', m3u8Settings_audioPids - Packet Identifier (PID) of the elementary audio stream(s) in the transport stream. Multiple values are accepted, and can be entered in ranges and/or by comma separation.

$sel:nielsenId3:M3u8Settings', m3u8Settings_nielsenId3 - If INSERT, Nielsen inaudible tones for media tracking will be detected in the input audio and an equivalent ID3 tag will be inserted in the output.

$sel:dataPTSControl:M3u8Settings', m3u8Settings_dataPTSControl - If you select ALIGN_TO_VIDEO, MediaConvert writes captions and data packets with Presentation Timestamp (PTS) values greater than or equal to the first video packet PTS (MediaConvert drops captions and data packets with lesser PTS values). Keep the default value (AUTO) to allow all PTS values.

$sel:pcrControl:M3u8Settings', m3u8Settings_pcrControl - When set to PCR_EVERY_PES_PACKET a Program Clock Reference value is inserted for every Packetized Elementary Stream (PES) header. This parameter is effective only when the PCR PID is the same as the video or audio elementary stream.

m3u8Settings_pmtPid :: Lens' M3u8Settings (Maybe Natural)

Packet Identifier (PID) for the Program Map Table (PMT) in the transport stream.

m3u8Settings_videoPid :: Lens' M3u8Settings (Maybe Natural)

Packet Identifier (PID) of the elementary video stream in the transport stream.

m3u8Settings_programNumber :: Lens' M3u8Settings (Maybe Natural)

The value of the program number field in the Program Map Table.

m3u8Settings_scte35Pid :: Lens' M3u8Settings (Maybe Natural)

Packet Identifier (PID) of the SCTE-35 stream in the transport stream.

m3u8Settings_transportStreamId :: Lens' M3u8Settings (Maybe Natural)

The value of the transport stream ID field in the Program Map Table.

m3u8Settings_maxPcrInterval :: Lens' M3u8Settings (Maybe Natural)

Specify the maximum time, in milliseconds, between Program Clock References (PCRs) inserted into the transport stream.

m3u8Settings_privateMetadataPid :: Lens' M3u8Settings (Maybe Natural)

Packet Identifier (PID) of the private metadata stream in the transport stream.

m3u8Settings_audioDuration :: Lens' M3u8Settings (Maybe M3u8AudioDuration)

Specify this setting only when your output will be consumed by a downstream repackaging workflow that is sensitive to very small duration differences between video and audio. For this situation, choose Match video duration (MATCH_VIDEO_DURATION). In all other cases, keep the default value, Default codec duration (DEFAULT_CODEC_DURATION). When you choose Match video duration, MediaConvert pads the output audio streams with silence or trims them to ensure that the total duration of each audio stream is at least as long as the total duration of the video stream. After padding or trimming, the audio stream duration is no more than one frame longer than the video stream. MediaConvert applies audio padding or trimming only to the end of the last segment of the output. For unsegmented outputs, MediaConvert adds padding only to the end of the file. When you keep the default value, any minor discrepancies between audio and video duration will depend on your output audio codec.

m3u8Settings_pmtInterval :: Lens' M3u8Settings (Maybe Natural)

The number of milliseconds between instances of this table in the output transport stream.

m3u8Settings_timedMetadataPid :: Lens' M3u8Settings (Maybe Natural)

Packet Identifier (PID) of the timed metadata stream in the transport stream.

m3u8Settings_audioFramesPerPes :: Lens' M3u8Settings (Maybe Natural)

The number of audio frames to insert for each PES packet.

m3u8Settings_pcrPid :: Lens' M3u8Settings (Maybe Natural)

Packet Identifier (PID) of the Program Clock Reference (PCR) in the transport stream. When no value is given, the encoder will assign the same value as the Video PID.

m3u8Settings_timedMetadata :: Lens' M3u8Settings (Maybe TimedMetadata)

Applies only to HLS outputs. Use this setting to specify whether the service inserts the ID3 timed metadata from the input in this output.

m3u8Settings_scte35Source :: Lens' M3u8Settings (Maybe M3u8Scte35Source)

For SCTE-35 markers from your input-- Choose Passthrough (PASSTHROUGH) if you want SCTE-35 markers that appear in your input to also appear in this output. Choose None (NONE) if you don't want SCTE-35 markers in this output. For SCTE-35 markers from an ESAM XML document-- Choose None (NONE) if you don't want manifest conditioning. Choose Passthrough (PASSTHROUGH) and choose Ad markers (adMarkers) if you do want manifest conditioning. In both cases, also provide the ESAM XML as a string in the setting Signal processing notification XML (sccXml).

m3u8Settings_patInterval :: Lens' M3u8Settings (Maybe Natural)

The number of milliseconds between instances of this table in the output transport stream.

m3u8Settings_audioPids :: Lens' M3u8Settings (Maybe [Natural])

Packet Identifier (PID) of the elementary audio stream(s) in the transport stream. Multiple values are accepted, and can be entered in ranges and/or by comma separation.

m3u8Settings_nielsenId3 :: Lens' M3u8Settings (Maybe M3u8NielsenId3)

If INSERT, Nielsen inaudible tones for media tracking will be detected in the input audio and an equivalent ID3 tag will be inserted in the output.

m3u8Settings_dataPTSControl :: Lens' M3u8Settings (Maybe M3u8DataPtsControl)

If you select ALIGN_TO_VIDEO, MediaConvert writes captions and data packets with Presentation Timestamp (PTS) values greater than or equal to the first video packet PTS (MediaConvert drops captions and data packets with lesser PTS values). Keep the default value (AUTO) to allow all PTS values.

m3u8Settings_pcrControl :: Lens' M3u8Settings (Maybe M3u8PcrControl)

When set to PCR_EVERY_PES_PACKET a Program Clock Reference value is inserted for every Packetized Elementary Stream (PES) header. This parameter is effective only when the PCR PID is the same as the video or audio elementary stream.