Amazonka.MediaConvert.Types.MotionImageInsertionFramerate

Description

Documentation

data MotionImageInsertionFramerate

For motion overlays that don't have a built-in frame rate, specify the frame rate of the overlay in frames per second, as a fraction. For example, specify 24 fps as 24/1. The overlay frame rate doesn't need to match the frame rate of the underlying video.

See: newMotionImageInsertionFramerate smart constructor.

Constructors

MotionImageInsertionFramerate'

Fields

framerateDenominator :: Maybe Natural

The bottom of the fraction that expresses your overlay frame rate. For example, if your frame rate is 24 fps, set this value to 1.

framerateNumerator :: Maybe Natural

The top of the fraction that expresses your overlay frame rate. For example, if your frame rate is 24 fps, set this value to 24.

newMotionImageInsertionFramerate :: MotionImageInsertionFramerate

Create a value of MotionImageInsertionFramerate with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:framerateDenominator:MotionImageInsertionFramerate', motionImageInsertionFramerate_framerateDenominator - The bottom of the fraction that expresses your overlay frame rate. For example, if your frame rate is 24 fps, set this value to 1.

$sel:framerateNumerator:MotionImageInsertionFramerate', motionImageInsertionFramerate_framerateNumerator - The top of the fraction that expresses your overlay frame rate. For example, if your frame rate is 24 fps, set this value to 24.

motionImageInsertionFramerate_framerateDenominator :: Lens' MotionImageInsertionFramerate (Maybe Natural)

The bottom of the fraction that expresses your overlay frame rate. For example, if your frame rate is 24 fps, set this value to 1.

motionImageInsertionFramerate_framerateNumerator :: Lens' MotionImageInsertionFramerate (Maybe Natural)

The top of the fraction that expresses your overlay frame rate. For example, if your frame rate is 24 fps, set this value to 24.