Amazonka.MediaConvert.Types.MotionImageInsertionOffset

Description

Documentation

data MotionImageInsertionOffset

Specify the offset between the upper-left corner of the video frame and the top left corner of the overlay.

See: newMotionImageInsertionOffset smart constructor.

Constructors

MotionImageInsertionOffset'

Fields

imageX :: Maybe Natural

Set the distance, in pixels, between the overlay and the left edge of the video frame.

imageY :: Maybe Natural

Set the distance, in pixels, between the overlay and the top edge of the video frame.

newMotionImageInsertionOffset :: MotionImageInsertionOffset

Create a value of MotionImageInsertionOffset with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:imageX:MotionImageInsertionOffset', motionImageInsertionOffset_imageX - Set the distance, in pixels, between the overlay and the left edge of the video frame.

$sel:imageY:MotionImageInsertionOffset', motionImageInsertionOffset_imageY - Set the distance, in pixels, between the overlay and the top edge of the video frame.

motionImageInsertionOffset_imageX :: Lens' MotionImageInsertionOffset (Maybe Natural)

Set the distance, in pixels, between the overlay and the left edge of the video frame.

motionImageInsertionOffset_imageY :: Lens' MotionImageInsertionOffset (Maybe Natural)

Set the distance, in pixels, between the overlay and the top edge of the video frame.