Amazonka.MediaConvert.Types.Preset

Description

Documentation

data Preset

A preset is a collection of preconfigured media conversion settings that you want MediaConvert to apply to the output during the conversion process.

See: newPreset smart constructor.

Constructors

Preset'

Fields

lastUpdated :: Maybe POSIX

The timestamp in epoch seconds when the preset was last updated.

arn :: Maybe Text

An identifier for this resource that is unique within all of AWS.

createdAt :: Maybe POSIX

The timestamp in epoch seconds for preset creation.

category :: Maybe Text

An optional category you create to organize your presets.

type' :: Maybe Type

A preset can be of two types: system or custom. System or built-in preset can't be modified or deleted by the user.

description :: Maybe Text

An optional description you create for each preset.

settings :: PresetSettings

Settings for preset

name :: Text

A name you create for each preset. Each name must be unique within your account.

newPreset

Arguments

:: PresetSettings
$sel:settings:Preset'

-> Text
$sel:name:Preset'

-> Preset

Create a value of Preset with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:lastUpdated:Preset', preset_lastUpdated - The timestamp in epoch seconds when the preset was last updated.

$sel:arn:Preset', preset_arn - An identifier for this resource that is unique within all of AWS.

$sel:createdAt:Preset', preset_createdAt - The timestamp in epoch seconds for preset creation.

$sel:category:Preset', preset_category - An optional category you create to organize your presets.

$sel:type':Preset', preset_type - A preset can be of two types: system or custom. System or built-in preset can't be modified or deleted by the user.

$sel:description:Preset', preset_description - An optional description you create for each preset.

$sel:settings:Preset', preset_settings - Settings for preset

$sel:name:Preset', preset_name - A name you create for each preset. Each name must be unique within your account.

preset_lastUpdated :: Lens' Preset (Maybe UTCTime)

The timestamp in epoch seconds when the preset was last updated.

preset_arn :: Lens' Preset (Maybe Text)

An identifier for this resource that is unique within all of AWS.

preset_createdAt :: Lens' Preset (Maybe UTCTime)

The timestamp in epoch seconds for preset creation.

preset_category :: Lens' Preset (Maybe Text)

An optional category you create to organize your presets.

preset_type :: Lens' Preset (Maybe Type)

A preset can be of two types: system or custom. System or built-in preset can't be modified or deleted by the user.

preset_description :: Lens' Preset (Maybe Text)

An optional description you create for each preset.

preset_settings :: Lens' Preset PresetSettings

Settings for preset

preset_name :: Lens' Preset Text

A name you create for each preset. Each name must be unique within your account.