Amazonka.MediaConvert.Types.QueueTransition

Description

Documentation

data QueueTransition

Description of the source and destination queues between which the job has moved, along with the timestamp of the move

See: newQueueTransition smart constructor.

Constructors

QueueTransition'

Fields

sourceQueue :: Maybe Text

The queue that the job was on before the transition.

destinationQueue :: Maybe Text

The queue that the job was on after the transition.

timestamp :: Maybe POSIX

The time, in Unix epoch format, that the job moved from the source queue to the destination queue.

newQueueTransition :: QueueTransition

Create a value of QueueTransition with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:sourceQueue:QueueTransition', queueTransition_sourceQueue - The queue that the job was on before the transition.

$sel:destinationQueue:QueueTransition', queueTransition_destinationQueue - The queue that the job was on after the transition.

$sel:timestamp:QueueTransition', queueTransition_timestamp - The time, in Unix epoch format, that the job moved from the source queue to the destination queue.

queueTransition_sourceQueue :: Lens' QueueTransition (Maybe Text)

The queue that the job was on before the transition.

queueTransition_destinationQueue :: Lens' QueueTransition (Maybe Text)

The queue that the job was on after the transition.

queueTransition_timestamp :: Lens' QueueTransition (Maybe UTCTime)

The time, in Unix epoch format, that the job moved from the source queue to the destination queue.