Amazonka.MediaConvert.Types.SpekeKeyProviderCmaf

Description

Documentation

data SpekeKeyProviderCmaf

If your output group type is CMAF, use these settings when doing DRM encryption with a SPEKE-compliant key provider. If your output group type is HLS, DASH, or Microsoft Smooth, use the SpekeKeyProvider settings instead.

See: newSpekeKeyProviderCmaf smart constructor.

Constructors

SpekeKeyProviderCmaf'

Fields

resourceId :: Maybe Text

Specify the resource ID that your SPEKE-compliant key provider uses to identify this content.

dashSignaledSystemIds :: Maybe [Text]

Specify the DRM system IDs that you want signaled in the DASH manifest that MediaConvert creates as part of this CMAF package. The DASH manifest can currently signal up to three system IDs. For more information, see https://dashif.org/identifiers/content_protection/.

certificateArn :: Maybe Text

If you want your key provider to encrypt the content keys that it provides to MediaConvert, set up a certificate with a master key using AWS Certificate Manager. Specify the certificate's Amazon Resource Name (ARN) here.

url :: Maybe Text

Specify the URL to the key server that your SPEKE-compliant DRM key provider uses to provide keys for encrypting your content.

hlsSignaledSystemIds :: Maybe [Text]

Specify the DRM system ID that you want signaled in the HLS manifest that MediaConvert creates as part of this CMAF package. The HLS manifest can currently signal only one system ID. For more information, see https://dashif.org/identifiers/content_protection/.

newSpekeKeyProviderCmaf :: SpekeKeyProviderCmaf

Create a value of SpekeKeyProviderCmaf with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:resourceId:SpekeKeyProviderCmaf', spekeKeyProviderCmaf_resourceId - Specify the resource ID that your SPEKE-compliant key provider uses to identify this content.

$sel:dashSignaledSystemIds:SpekeKeyProviderCmaf', spekeKeyProviderCmaf_dashSignaledSystemIds - Specify the DRM system IDs that you want signaled in the DASH manifest that MediaConvert creates as part of this CMAF package. The DASH manifest can currently signal up to three system IDs. For more information, see https://dashif.org/identifiers/content_protection/.

$sel:certificateArn:SpekeKeyProviderCmaf', spekeKeyProviderCmaf_certificateArn - If you want your key provider to encrypt the content keys that it provides to MediaConvert, set up a certificate with a master key using AWS Certificate Manager. Specify the certificate's Amazon Resource Name (ARN) here.

$sel:url:SpekeKeyProviderCmaf', spekeKeyProviderCmaf_url - Specify the URL to the key server that your SPEKE-compliant DRM key provider uses to provide keys for encrypting your content.

$sel:hlsSignaledSystemIds:SpekeKeyProviderCmaf', spekeKeyProviderCmaf_hlsSignaledSystemIds - Specify the DRM system ID that you want signaled in the HLS manifest that MediaConvert creates as part of this CMAF package. The HLS manifest can currently signal only one system ID. For more information, see https://dashif.org/identifiers/content_protection/.

spekeKeyProviderCmaf_resourceId :: Lens' SpekeKeyProviderCmaf (Maybe Text)

Specify the resource ID that your SPEKE-compliant key provider uses to identify this content.

spekeKeyProviderCmaf_dashSignaledSystemIds :: Lens' SpekeKeyProviderCmaf (Maybe [Text])

Specify the DRM system IDs that you want signaled in the DASH manifest that MediaConvert creates as part of this CMAF package. The DASH manifest can currently signal up to three system IDs. For more information, see https://dashif.org/identifiers/content_protection/.

spekeKeyProviderCmaf_certificateArn :: Lens' SpekeKeyProviderCmaf (Maybe Text)

If you want your key provider to encrypt the content keys that it provides to MediaConvert, set up a certificate with a master key using AWS Certificate Manager. Specify the certificate's Amazon Resource Name (ARN) here.

spekeKeyProviderCmaf_url :: Lens' SpekeKeyProviderCmaf (Maybe Text)

Specify the URL to the key server that your SPEKE-compliant DRM key provider uses to provide keys for encrypting your content.

spekeKeyProviderCmaf_hlsSignaledSystemIds :: Lens' SpekeKeyProviderCmaf (Maybe [Text])

Specify the DRM system ID that you want signaled in the HLS manifest that MediaConvert creates as part of this CMAF package. The HLS manifest can currently signal only one system ID. For more information, see https://dashif.org/identifiers/content_protection/.