Amazonka.MediaConvert.Types.StaticKeyProvider

Description

Documentation

data StaticKeyProvider

Use these settings to set up encryption with a static key provider.

See: newStaticKeyProvider smart constructor.

Constructors

StaticKeyProvider'

Fields

staticKeyValue :: Maybe Text

Relates to DRM implementation. Use a 32-character hexidecimal string to specify Key Value (StaticKeyValue).

url :: Maybe Text

Relates to DRM implementation. The location of the license server used for protecting content.

keyFormat :: Maybe Text

Relates to DRM implementation. Sets the value of the KEYFORMAT attribute. Must be 'identity' or a reverse DNS string. May be omitted to indicate an implicit value of 'identity'.

keyFormatVersions :: Maybe Text

Relates to DRM implementation. Either a single positive integer version value or a slash delimited list of version values (1/2/3).

newStaticKeyProvider :: StaticKeyProvider

Create a value of StaticKeyProvider with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:staticKeyValue:StaticKeyProvider', staticKeyProvider_staticKeyValue - Relates to DRM implementation. Use a 32-character hexidecimal string to specify Key Value (StaticKeyValue).

$sel:url:StaticKeyProvider', staticKeyProvider_url - Relates to DRM implementation. The location of the license server used for protecting content.

$sel:keyFormat:StaticKeyProvider', staticKeyProvider_keyFormat - Relates to DRM implementation. Sets the value of the KEYFORMAT attribute. Must be 'identity' or a reverse DNS string. May be omitted to indicate an implicit value of 'identity'.

$sel:keyFormatVersions:StaticKeyProvider', staticKeyProvider_keyFormatVersions - Relates to DRM implementation. Either a single positive integer version value or a slash delimited list of version values (1/2/3).

staticKeyProvider_staticKeyValue :: Lens' StaticKeyProvider (Maybe Text)

Relates to DRM implementation. Use a 32-character hexidecimal string to specify Key Value (StaticKeyValue).

staticKeyProvider_url :: Lens' StaticKeyProvider (Maybe Text)

Relates to DRM implementation. The location of the license server used for protecting content.

staticKeyProvider_keyFormat :: Lens' StaticKeyProvider (Maybe Text)

Relates to DRM implementation. Sets the value of the KEYFORMAT attribute. Must be 'identity' or a reverse DNS string. May be omitted to indicate an implicit value of 'identity'.

staticKeyProvider_keyFormatVersions :: Lens' StaticKeyProvider (Maybe Text)

Relates to DRM implementation. Either a single positive integer version value or a slash delimited list of version values (1/2/3).