Amazonka.MediaConvert.Types.TimecodeBurnin

Description

Documentation

data TimecodeBurnin

Settings for burning the output timecode and specified prefix into the output.

See: newTimecodeBurnin smart constructor.

Constructors

TimecodeBurnin'

Fields

prefix :: Maybe Text

Use Prefix (Prefix) to place ASCII characters before any burned-in timecode. For example, a prefix of "EZ-" will result in the timecode "EZ-00:00:00:00". Provide either the characters themselves or the ASCII code equivalents. The supported range of characters is 0x20 through 0x7e. This includes letters, numbers, and all special characters represented on a standard English keyboard.

fontSize :: Maybe Natural

Use Font Size (FontSize) to set the font size of any burned-in timecode. Valid values are 10, 16, 32, 48.

position :: Maybe TimecodeBurninPosition

Use Position (Position) under under Timecode burn-in (TimecodeBurnIn) to specify the location the burned-in timecode on output video.

newTimecodeBurnin :: TimecodeBurnin

Create a value of TimecodeBurnin with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:prefix:TimecodeBurnin', timecodeBurnin_prefix - Use Prefix (Prefix) to place ASCII characters before any burned-in timecode. For example, a prefix of "EZ-" will result in the timecode "EZ-00:00:00:00". Provide either the characters themselves or the ASCII code equivalents. The supported range of characters is 0x20 through 0x7e. This includes letters, numbers, and all special characters represented on a standard English keyboard.

$sel:fontSize:TimecodeBurnin', timecodeBurnin_fontSize - Use Font Size (FontSize) to set the font size of any burned-in timecode. Valid values are 10, 16, 32, 48.

$sel:position:TimecodeBurnin', timecodeBurnin_position - Use Position (Position) under under Timecode burn-in (TimecodeBurnIn) to specify the location the burned-in timecode on output video.

timecodeBurnin_prefix :: Lens' TimecodeBurnin (Maybe Text)

Use Prefix (Prefix) to place ASCII characters before any burned-in timecode. For example, a prefix of "EZ-" will result in the timecode "EZ-00:00:00:00". Provide either the characters themselves or the ASCII code equivalents. The supported range of characters is 0x20 through 0x7e. This includes letters, numbers, and all special characters represented on a standard English keyboard.

timecodeBurnin_fontSize :: Lens' TimecodeBurnin (Maybe Natural)

Use Font Size (FontSize) to set the font size of any burned-in timecode. Valid values are 10, 16, 32, 48.

timecodeBurnin_position :: Lens' TimecodeBurnin (Maybe TimecodeBurninPosition)

Use Position (Position) under under Timecode burn-in (TimecodeBurnIn) to specify the location the burned-in timecode on output video.