Amazonka.MediaConvert.Types.TimedMetadataInsertion

Description

Documentation

data TimedMetadataInsertion

Enable Timed metadata insertion (TimedMetadataInsertion) to include ID3 tags in any HLS outputs. To include timed metadata, you must enable it here, enable it in each output container, and specify tags and timecodes in ID3 insertion (Id3Insertion) objects.

See: newTimedMetadataInsertion smart constructor.

Constructors

TimedMetadataInsertion'

Fields

id3Insertions :: Maybe [Id3Insertion]

Id3Insertions contains the array of Id3Insertion instances.

newTimedMetadataInsertion :: TimedMetadataInsertion

Create a value of TimedMetadataInsertion with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:id3Insertions:TimedMetadataInsertion', timedMetadataInsertion_id3Insertions - Id3Insertions contains the array of Id3Insertion instances.

timedMetadataInsertion_id3Insertions :: Lens' TimedMetadataInsertion (Maybe [Id3Insertion])

Id3Insertions contains the array of Id3Insertion instances.