Amazonka.MediaConvert.Types.VideoPreprocessor

Description

Documentation

data VideoPreprocessor

Find additional transcoding features under Preprocessors (VideoPreprocessors). Enable the features at each output individually. These features are disabled by default.

See: newVideoPreprocessor smart constructor.

Constructors

VideoPreprocessor'

Fields

timecodeBurnin :: Maybe TimecodeBurnin

Settings for burning the output timecode and specified prefix into the output.

dolbyVision :: Maybe DolbyVision

Enable Dolby Vision feature to produce Dolby Vision compatible video output.

colorCorrector :: Maybe ColorCorrector

Use these settings to convert the color space or to modify properties such as hue and contrast for this output. For more information, see https://docs.aws.amazon.com/mediaconvert/latest/ug/converting-the-color-space.html.

deinterlacer :: Maybe Deinterlacer

Use the deinterlacer to produce smoother motion and a clearer picture. For more information, see https://docs.aws.amazon.com/mediaconvert/latest/ug/working-with-scan-type.html.

noiseReducer :: Maybe NoiseReducer

Enable the Noise reducer (NoiseReducer) feature to remove noise from your video output if necessary. Enable or disable this feature for each output individually. This setting is disabled by default.

imageInserter :: Maybe ImageInserter

Enable the Image inserter (ImageInserter) feature to include a graphic overlay on your video. Enable or disable this feature for each output individually. This setting is disabled by default.

hdr10Plus :: Maybe Hdr10Plus

Enable HDR10+ analyis and metadata injection. Compatible with HEVC only.

partnerWatermarking :: Maybe PartnerWatermarking

If you work with a third party video watermarking partner, use the group of settings that correspond with your watermarking partner to include watermarks in your output.

newVideoPreprocessor :: VideoPreprocessor

Create a value of VideoPreprocessor with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:timecodeBurnin:VideoPreprocessor', videoPreprocessor_timecodeBurnin - Settings for burning the output timecode and specified prefix into the output.

$sel:dolbyVision:VideoPreprocessor', videoPreprocessor_dolbyVision - Enable Dolby Vision feature to produce Dolby Vision compatible video output.

$sel:colorCorrector:VideoPreprocessor', videoPreprocessor_colorCorrector - Use these settings to convert the color space or to modify properties such as hue and contrast for this output. For more information, see https://docs.aws.amazon.com/mediaconvert/latest/ug/converting-the-color-space.html.

$sel:deinterlacer:VideoPreprocessor', videoPreprocessor_deinterlacer - Use the deinterlacer to produce smoother motion and a clearer picture. For more information, see https://docs.aws.amazon.com/mediaconvert/latest/ug/working-with-scan-type.html.

$sel:noiseReducer:VideoPreprocessor', videoPreprocessor_noiseReducer - Enable the Noise reducer (NoiseReducer) feature to remove noise from your video output if necessary. Enable or disable this feature for each output individually. This setting is disabled by default.

$sel:imageInserter:VideoPreprocessor', videoPreprocessor_imageInserter - Enable the Image inserter (ImageInserter) feature to include a graphic overlay on your video. Enable or disable this feature for each output individually. This setting is disabled by default.

$sel:hdr10Plus:VideoPreprocessor', videoPreprocessor_hdr10Plus - Enable HDR10+ analyis and metadata injection. Compatible with HEVC only.

$sel:partnerWatermarking:VideoPreprocessor', videoPreprocessor_partnerWatermarking - If you work with a third party video watermarking partner, use the group of settings that correspond with your watermarking partner to include watermarks in your output.

videoPreprocessor_timecodeBurnin :: Lens' VideoPreprocessor (Maybe TimecodeBurnin)

Settings for burning the output timecode and specified prefix into the output.

videoPreprocessor_dolbyVision :: Lens' VideoPreprocessor (Maybe DolbyVision)

Enable Dolby Vision feature to produce Dolby Vision compatible video output.

videoPreprocessor_colorCorrector :: Lens' VideoPreprocessor (Maybe ColorCorrector)

Use these settings to convert the color space or to modify properties such as hue and contrast for this output. For more information, see https://docs.aws.amazon.com/mediaconvert/latest/ug/converting-the-color-space.html.

videoPreprocessor_deinterlacer :: Lens' VideoPreprocessor (Maybe Deinterlacer)

Use the deinterlacer to produce smoother motion and a clearer picture. For more information, see https://docs.aws.amazon.com/mediaconvert/latest/ug/working-with-scan-type.html.

videoPreprocessor_noiseReducer :: Lens' VideoPreprocessor (Maybe NoiseReducer)

Enable the Noise reducer (NoiseReducer) feature to remove noise from your video output if necessary. Enable or disable this feature for each output individually. This setting is disabled by default.

videoPreprocessor_imageInserter :: Lens' VideoPreprocessor (Maybe ImageInserter)

Enable the Image inserter (ImageInserter) feature to include a graphic overlay on your video. Enable or disable this feature for each output individually. This setting is disabled by default.

videoPreprocessor_hdr10Plus :: Lens' VideoPreprocessor (Maybe Hdr10Plus)

Enable HDR10+ analyis and metadata injection. Compatible with HEVC only.

videoPreprocessor_partnerWatermarking :: Lens' VideoPreprocessor (Maybe PartnerWatermarking)

If you work with a third party video watermarking partner, use the group of settings that correspond with your watermarking partner to include watermarks in your output.