Amazonka.MediaLive.DeleteMultiplex

Description

Delete a multiplex. The multiplex must be idle.

Creating a Request

data DeleteMultiplex

Placeholder documentation for DeleteMultiplexRequest

See: newDeleteMultiplex smart constructor.

Constructors

DeleteMultiplex'

Fields

multiplexId :: Text

The ID of the multiplex.

newDeleteMultiplex

Arguments

:: Text
$sel:multiplexId:DeleteMultiplex'

-> DeleteMultiplex

Create a value of DeleteMultiplex with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:multiplexId:DeleteMultiplex', deleteMultiplex_multiplexId - The ID of the multiplex.

Request Lenses

deleteMultiplex_multiplexId :: Lens' DeleteMultiplex Text

The ID of the multiplex.

Destructuring the Response

data DeleteMultiplexResponse

Placeholder documentation for DeleteMultiplexResponse

See: newDeleteMultiplexResponse smart constructor.

Constructors

DeleteMultiplexResponse'

Fields

state :: Maybe MultiplexState

The current state of the multiplex.

arn :: Maybe Text

The unique arn of the multiplex.

pipelinesRunningCount :: Maybe Int

The number of currently healthy pipelines.

availabilityZones :: Maybe [Text]

A list of availability zones for the multiplex.

programCount :: Maybe Int

The number of programs in the multiplex.

destinations :: Maybe [MultiplexOutputDestination]

A list of the multiplex output destinations.

name :: Maybe Text

The name of the multiplex.

id :: Maybe Text

The unique id of the multiplex.

multiplexSettings :: Maybe MultiplexSettings

Configuration for a multiplex event.

tags :: Maybe (HashMap Text Text)

A collection of key-value pairs.

httpStatus :: Int

The response's http status code.

newDeleteMultiplexResponse

Arguments

:: Int
$sel:httpStatus:DeleteMultiplexResponse'

-> DeleteMultiplexResponse

Create a value of DeleteMultiplexResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:state:DeleteMultiplexResponse', deleteMultiplexResponse_state - The current state of the multiplex.

$sel:arn:DeleteMultiplexResponse', deleteMultiplexResponse_arn - The unique arn of the multiplex.

$sel:pipelinesRunningCount:DeleteMultiplexResponse', deleteMultiplexResponse_pipelinesRunningCount - The number of currently healthy pipelines.

$sel:availabilityZones:DeleteMultiplexResponse', deleteMultiplexResponse_availabilityZones - A list of availability zones for the multiplex.

$sel:programCount:DeleteMultiplexResponse', deleteMultiplexResponse_programCount - The number of programs in the multiplex.

$sel:destinations:DeleteMultiplexResponse', deleteMultiplexResponse_destinations - A list of the multiplex output destinations.

$sel:name:DeleteMultiplexResponse', deleteMultiplexResponse_name - The name of the multiplex.

$sel:id:DeleteMultiplexResponse', deleteMultiplexResponse_id - The unique id of the multiplex.

$sel:multiplexSettings:DeleteMultiplexResponse', deleteMultiplexResponse_multiplexSettings - Configuration for a multiplex event.

$sel:tags:DeleteMultiplexResponse', deleteMultiplexResponse_tags - A collection of key-value pairs.

$sel:httpStatus:DeleteMultiplexResponse', deleteMultiplexResponse_httpStatus - The response's http status code.

Response Lenses

deleteMultiplexResponse_state :: Lens' DeleteMultiplexResponse (Maybe MultiplexState)

The current state of the multiplex.

deleteMultiplexResponse_arn :: Lens' DeleteMultiplexResponse (Maybe Text)

The unique arn of the multiplex.

deleteMultiplexResponse_pipelinesRunningCount :: Lens' DeleteMultiplexResponse (Maybe Int)

The number of currently healthy pipelines.

deleteMultiplexResponse_availabilityZones :: Lens' DeleteMultiplexResponse (Maybe [Text])

A list of availability zones for the multiplex.

deleteMultiplexResponse_programCount :: Lens' DeleteMultiplexResponse (Maybe Int)

The number of programs in the multiplex.

deleteMultiplexResponse_destinations :: Lens' DeleteMultiplexResponse (Maybe [MultiplexOutputDestination])

A list of the multiplex output destinations.

deleteMultiplexResponse_name :: Lens' DeleteMultiplexResponse (Maybe Text)

The name of the multiplex.

deleteMultiplexResponse_id :: Lens' DeleteMultiplexResponse (Maybe Text)

The unique id of the multiplex.

deleteMultiplexResponse_multiplexSettings :: Lens' DeleteMultiplexResponse (Maybe MultiplexSettings)

Configuration for a multiplex event.

deleteMultiplexResponse_tags :: Lens' DeleteMultiplexResponse (Maybe (HashMap Text Text))

A collection of key-value pairs.

deleteMultiplexResponse_httpStatus :: Lens' DeleteMultiplexResponse Int

The response's http status code.