Amazonka.MediaLive.DeleteMultiplexProgram

Description

Delete a program from a multiplex.

Creating a Request

data DeleteMultiplexProgram

Placeholder documentation for DeleteMultiplexProgramRequest

See: newDeleteMultiplexProgram smart constructor.

Constructors

DeleteMultiplexProgram'

Fields

multiplexId :: Text

The ID of the multiplex that the program belongs to.

programName :: Text

The multiplex program name.

newDeleteMultiplexProgram

Arguments

:: Text
$sel:multiplexId:DeleteMultiplexProgram'

-> Text
$sel:programName:DeleteMultiplexProgram'

-> DeleteMultiplexProgram

Create a value of DeleteMultiplexProgram with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:multiplexId:DeleteMultiplexProgram', deleteMultiplexProgram_multiplexId - The ID of the multiplex that the program belongs to.

$sel:programName:DeleteMultiplexProgram', deleteMultiplexProgram_programName - The multiplex program name.

Request Lenses

deleteMultiplexProgram_multiplexId :: Lens' DeleteMultiplexProgram Text

The ID of the multiplex that the program belongs to.

deleteMultiplexProgram_programName :: Lens' DeleteMultiplexProgram Text

The multiplex program name.

Destructuring the Response

data DeleteMultiplexProgramResponse

Placeholder documentation for DeleteMultiplexProgramResponse

See: newDeleteMultiplexProgramResponse smart constructor.

Constructors

DeleteMultiplexProgramResponse'

Fields

packetIdentifiersMap :: Maybe MultiplexProgramPacketIdentifiersMap

The packet identifier map for this multiplex program.

pipelineDetails :: Maybe [MultiplexProgramPipelineDetail]

Contains information about the current sources for the specified program in the specified multiplex. Keep in mind that each multiplex pipeline connects to both pipelines in a given source channel (the channel identified by the program). But only one of those channel pipelines is ever active at one time.

programName :: Maybe Text

The name of the multiplex program.

channelId :: Maybe Text

The MediaLive channel associated with the program.

multiplexProgramSettings :: Maybe MultiplexProgramSettings

The settings for this multiplex program.

httpStatus :: Int

The response's http status code.

newDeleteMultiplexProgramResponse

Arguments

:: Int
$sel:httpStatus:DeleteMultiplexProgramResponse'

-> DeleteMultiplexProgramResponse

Create a value of DeleteMultiplexProgramResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:packetIdentifiersMap:DeleteMultiplexProgramResponse', deleteMultiplexProgramResponse_packetIdentifiersMap - The packet identifier map for this multiplex program.

$sel:pipelineDetails:DeleteMultiplexProgramResponse', deleteMultiplexProgramResponse_pipelineDetails - Contains information about the current sources for the specified program in the specified multiplex. Keep in mind that each multiplex pipeline connects to both pipelines in a given source channel (the channel identified by the program). But only one of those channel pipelines is ever active at one time.

$sel:programName:DeleteMultiplexProgram', deleteMultiplexProgramResponse_programName - The name of the multiplex program.

$sel:channelId:DeleteMultiplexProgramResponse', deleteMultiplexProgramResponse_channelId - The MediaLive channel associated with the program.

$sel:multiplexProgramSettings:DeleteMultiplexProgramResponse', deleteMultiplexProgramResponse_multiplexProgramSettings - The settings for this multiplex program.

$sel:httpStatus:DeleteMultiplexProgramResponse', deleteMultiplexProgramResponse_httpStatus - The response's http status code.

Response Lenses

deleteMultiplexProgramResponse_packetIdentifiersMap :: Lens' DeleteMultiplexProgramResponse (Maybe MultiplexProgramPacketIdentifiersMap)

The packet identifier map for this multiplex program.

deleteMultiplexProgramResponse_pipelineDetails :: Lens' DeleteMultiplexProgramResponse (Maybe [MultiplexProgramPipelineDetail])

Contains information about the current sources for the specified program in the specified multiplex. Keep in mind that each multiplex pipeline connects to both pipelines in a given source channel (the channel identified by the program). But only one of those channel pipelines is ever active at one time.

deleteMultiplexProgramResponse_programName :: Lens' DeleteMultiplexProgramResponse (Maybe Text)

The name of the multiplex program.

deleteMultiplexProgramResponse_channelId :: Lens' DeleteMultiplexProgramResponse (Maybe Text)

The MediaLive channel associated with the program.

deleteMultiplexProgramResponse_multiplexProgramSettings :: Lens' DeleteMultiplexProgramResponse (Maybe MultiplexProgramSettings)

The settings for this multiplex program.

deleteMultiplexProgramResponse_httpStatus :: Lens' DeleteMultiplexProgramResponse Int

The response's http status code.