Amazonka.MediaLive.Types.DvbTdtSettings

Description

Documentation

data DvbTdtSettings

DVB Time and Date Table (SDT)

See: newDvbTdtSettings smart constructor.

Constructors

DvbTdtSettings'

Fields

repInterval :: Maybe Natural

The number of milliseconds between instances of this table in the output transport stream.

newDvbTdtSettings :: DvbTdtSettings

Create a value of DvbTdtSettings with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:repInterval:DvbTdtSettings', dvbTdtSettings_repInterval - The number of milliseconds between instances of this table in the output transport stream.

dvbTdtSettings_repInterval :: Lens' DvbTdtSettings (Maybe Natural)

The number of milliseconds between instances of this table in the output transport stream.