Amazonka.MediaLive.Types.HlsOutputSettings

Description

Documentation

data HlsOutputSettings

Hls Output Settings

See: newHlsOutputSettings smart constructor.

Constructors

HlsOutputSettings'

Fields

h265PackagingType :: Maybe HlsH265PackagingType

Only applicable when this output is referencing an H.265 video description. Specifies whether MP4 segments should be packaged as HEV1 or HVC1.

segmentModifier :: Maybe Text

String concatenated to end of segment filenames.

nameModifier :: Maybe Text

String concatenated to the end of the destination filename. Accepts \"Format Identifiers\":#formatIdentifierParameters.

hlsSettings :: HlsSettings

Settings regarding the underlying stream. These settings are different for audio-only outputs.

newHlsOutputSettings

Arguments

:: HlsSettings
$sel:hlsSettings:HlsOutputSettings'

-> HlsOutputSettings

Create a value of HlsOutputSettings with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:h265PackagingType:HlsOutputSettings', hlsOutputSettings_h265PackagingType - Only applicable when this output is referencing an H.265 video description. Specifies whether MP4 segments should be packaged as HEV1 or HVC1.

$sel:segmentModifier:HlsOutputSettings', hlsOutputSettings_segmentModifier - String concatenated to end of segment filenames.

$sel:nameModifier:HlsOutputSettings', hlsOutputSettings_nameModifier - String concatenated to the end of the destination filename. Accepts \"Format Identifiers\":#formatIdentifierParameters.

$sel:hlsSettings:HlsOutputSettings', hlsOutputSettings_hlsSettings - Settings regarding the underlying stream. These settings are different for audio-only outputs.

hlsOutputSettings_h265PackagingType :: Lens' HlsOutputSettings (Maybe HlsH265PackagingType)

Only applicable when this output is referencing an H.265 video description. Specifies whether MP4 segments should be packaged as HEV1 or HVC1.

hlsOutputSettings_segmentModifier :: Lens' HlsOutputSettings (Maybe Text)

String concatenated to end of segment filenames.

hlsOutputSettings_nameModifier :: Lens' HlsOutputSettings (Maybe Text)

String concatenated to the end of the destination filename. Accepts \"Format Identifiers\":#formatIdentifierParameters.

hlsOutputSettings_hlsSettings :: Lens' HlsOutputSettings HlsSettings

Settings regarding the underlying stream. These settings are different for audio-only outputs.