Amazonka.MediaLive.Types.InputClippingSettings

Description

Documentation

data InputClippingSettings

Settings to let you create a clip of the file input, in order to set up the input to ingest only a portion of the file.

See: newInputClippingSettings smart constructor.

Constructors

InputClippingSettings'

Fields

stopTimecode :: Maybe StopTimecode

Settings to identify the end of the clip.

startTimecode :: Maybe StartTimecode

Settings to identify the start of the clip.

inputTimecodeSource :: InputTimecodeSource

The source of the timecodes in the source being clipped.

newInputClippingSettings

Arguments

:: InputTimecodeSource
$sel:inputTimecodeSource:InputClippingSettings'

-> InputClippingSettings

Create a value of InputClippingSettings with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:stopTimecode:InputClippingSettings', inputClippingSettings_stopTimecode - Settings to identify the end of the clip.

$sel:startTimecode:InputClippingSettings', inputClippingSettings_startTimecode - Settings to identify the start of the clip.

$sel:inputTimecodeSource:InputClippingSettings', inputClippingSettings_inputTimecodeSource - The source of the timecodes in the source being clipped.

inputClippingSettings_stopTimecode :: Lens' InputClippingSettings (Maybe StopTimecode)

Settings to identify the end of the clip.

inputClippingSettings_startTimecode :: Lens' InputClippingSettings (Maybe StartTimecode)

Settings to identify the start of the clip.

inputClippingSettings_inputTimecodeSource :: Lens' InputClippingSettings InputTimecodeSource

The source of the timecodes in the source being clipped.