Amazonka.MediaLive.Types.InputDestination

Description

Documentation

data InputDestination

The settings for a PUSH type input.

See: newInputDestination smart constructor.

Constructors

InputDestination'

Fields

url :: Maybe Text

This represents the endpoint that the customer stream will be pushed to.

ip :: Maybe Text

The system-generated static IP address of endpoint. It remains fixed for the lifetime of the input.

vpc :: Maybe InputDestinationVpc

port :: Maybe Text

The port number for the input.

newInputDestination :: InputDestination

Create a value of InputDestination with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:url:InputDestination', inputDestination_url - This represents the endpoint that the customer stream will be pushed to.

$sel:ip:InputDestination', inputDestination_ip - The system-generated static IP address of endpoint. It remains fixed for the lifetime of the input.

$sel:vpc:InputDestination', inputDestination_vpc - Undocumented member.

$sel:port:InputDestination', inputDestination_port - The port number for the input.

inputDestination_url :: Lens' InputDestination (Maybe Text)

This represents the endpoint that the customer stream will be pushed to.

inputDestination_ip :: Lens' InputDestination (Maybe Text)

The system-generated static IP address of endpoint. It remains fixed for the lifetime of the input.

inputDestination_vpc :: Lens' InputDestination (Maybe InputDestinationVpc)

Undocumented member.

inputDestination_port :: Lens' InputDestination (Maybe Text)

The port number for the input.