Amazonka.MediaLive.Types.InputDeviceSummary

Description

Documentation

data InputDeviceSummary

Details of the input device.

See: newInputDeviceSummary smart constructor.

Constructors

InputDeviceSummary'

Fields

arn :: Maybe Text

The unique ARN of the input device.

macAddress :: Maybe Text

The network MAC address of the input device.

hdDeviceSettings :: Maybe InputDeviceHdSettings

Settings that describe an input device that is type HD.

uhdDeviceSettings :: Maybe InputDeviceUhdSettings

Settings that describe an input device that is type UHD.

name :: Maybe Text

A name that you specify for the input device.

id :: Maybe Text

The unique ID of the input device.

deviceUpdateStatus :: Maybe DeviceUpdateStatus

The status of software on the input device.

deviceSettingsSyncState :: Maybe DeviceSettingsSyncState

The status of the action to synchronize the device configuration. If you change the configuration of the input device (for example, the maximum bitrate), MediaLive sends the new data to the device. The device might not update itself immediately. SYNCED means the device has updated its configuration. SYNCING means that it has not updated its configuration.

type' :: Maybe InputDeviceType

The type of the input device.

serialNumber :: Maybe Text

The unique serial number of the input device.

networkSettings :: Maybe InputDeviceNetworkSettings

Network settings for the input device.

connectionState :: Maybe InputDeviceConnectionState

The state of the connection between the input device and AWS.

newInputDeviceSummary :: InputDeviceSummary

Create a value of InputDeviceSummary with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:arn:InputDeviceSummary', inputDeviceSummary_arn - The unique ARN of the input device.

$sel:macAddress:InputDeviceSummary', inputDeviceSummary_macAddress - The network MAC address of the input device.

$sel:hdDeviceSettings:InputDeviceSummary', inputDeviceSummary_hdDeviceSettings - Settings that describe an input device that is type HD.

$sel:uhdDeviceSettings:InputDeviceSummary', inputDeviceSummary_uhdDeviceSettings - Settings that describe an input device that is type UHD.

$sel:name:InputDeviceSummary', inputDeviceSummary_name - A name that you specify for the input device.

$sel:id:InputDeviceSummary', inputDeviceSummary_id - The unique ID of the input device.

$sel:deviceUpdateStatus:InputDeviceSummary', inputDeviceSummary_deviceUpdateStatus - The status of software on the input device.

$sel:deviceSettingsSyncState:InputDeviceSummary', inputDeviceSummary_deviceSettingsSyncState - The status of the action to synchronize the device configuration. If you change the configuration of the input device (for example, the maximum bitrate), MediaLive sends the new data to the device. The device might not update itself immediately. SYNCED means the device has updated its configuration. SYNCING means that it has not updated its configuration.

$sel:type':InputDeviceSummary', inputDeviceSummary_type - The type of the input device.

$sel:serialNumber:InputDeviceSummary', inputDeviceSummary_serialNumber - The unique serial number of the input device.

$sel:networkSettings:InputDeviceSummary', inputDeviceSummary_networkSettings - Network settings for the input device.

$sel:connectionState:InputDeviceSummary', inputDeviceSummary_connectionState - The state of the connection between the input device and AWS.

inputDeviceSummary_arn :: Lens' InputDeviceSummary (Maybe Text)

The unique ARN of the input device.

inputDeviceSummary_macAddress :: Lens' InputDeviceSummary (Maybe Text)

The network MAC address of the input device.

inputDeviceSummary_hdDeviceSettings :: Lens' InputDeviceSummary (Maybe InputDeviceHdSettings)

Settings that describe an input device that is type HD.

inputDeviceSummary_uhdDeviceSettings :: Lens' InputDeviceSummary (Maybe InputDeviceUhdSettings)

Settings that describe an input device that is type UHD.

inputDeviceSummary_name :: Lens' InputDeviceSummary (Maybe Text)

A name that you specify for the input device.

inputDeviceSummary_id :: Lens' InputDeviceSummary (Maybe Text)

The unique ID of the input device.

inputDeviceSummary_deviceUpdateStatus :: Lens' InputDeviceSummary (Maybe DeviceUpdateStatus)

The status of software on the input device.

inputDeviceSummary_deviceSettingsSyncState :: Lens' InputDeviceSummary (Maybe DeviceSettingsSyncState)

The status of the action to synchronize the device configuration. If you change the configuration of the input device (for example, the maximum bitrate), MediaLive sends the new data to the device. The device might not update itself immediately. SYNCED means the device has updated its configuration. SYNCING means that it has not updated its configuration.

inputDeviceSummary_type :: Lens' InputDeviceSummary (Maybe InputDeviceType)

The type of the input device.

inputDeviceSummary_serialNumber :: Lens' InputDeviceSummary (Maybe Text)

The unique serial number of the input device.

inputDeviceSummary_networkSettings :: Lens' InputDeviceSummary (Maybe InputDeviceNetworkSettings)

Network settings for the input device.

inputDeviceSummary_connectionState :: Lens' InputDeviceSummary (Maybe InputDeviceConnectionState)

The state of the connection between the input device and AWS.