Amazonka.MediaLive.Types.InputSecurityGroup

Description

Documentation

data InputSecurityGroup

An Input Security Group

See: newInputSecurityGroup smart constructor.

Constructors

InputSecurityGroup'

Fields

state :: Maybe InputSecurityGroupState

The current state of the Input Security Group.

arn :: Maybe Text

Unique ARN of Input Security Group

inputs :: Maybe [Text]

The list of inputs currently using this Input Security Group.

id :: Maybe Text

The Id of the Input Security Group

whitelistRules :: Maybe [InputWhitelistRule]

Whitelist rules and their sync status

tags :: Maybe (HashMap Text Text)

A collection of key-value pairs.

newInputSecurityGroup :: InputSecurityGroup

Create a value of InputSecurityGroup with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:state:InputSecurityGroup', inputSecurityGroup_state - The current state of the Input Security Group.

$sel:arn:InputSecurityGroup', inputSecurityGroup_arn - Unique ARN of Input Security Group

$sel:inputs:InputSecurityGroup', inputSecurityGroup_inputs - The list of inputs currently using this Input Security Group.

$sel:id:InputSecurityGroup', inputSecurityGroup_id - The Id of the Input Security Group

$sel:whitelistRules:InputSecurityGroup', inputSecurityGroup_whitelistRules - Whitelist rules and their sync status

$sel:tags:InputSecurityGroup', inputSecurityGroup_tags - A collection of key-value pairs.

inputSecurityGroup_state :: Lens' InputSecurityGroup (Maybe InputSecurityGroupState)

The current state of the Input Security Group.

inputSecurityGroup_arn :: Lens' InputSecurityGroup (Maybe Text)

Unique ARN of Input Security Group

inputSecurityGroup_inputs :: Lens' InputSecurityGroup (Maybe [Text])

The list of inputs currently using this Input Security Group.

inputSecurityGroup_id :: Lens' InputSecurityGroup (Maybe Text)

The Id of the Input Security Group

inputSecurityGroup_whitelistRules :: Lens' InputSecurityGroup (Maybe [InputWhitelistRule])

Whitelist rules and their sync status

inputSecurityGroup_tags :: Lens' InputSecurityGroup (Maybe (HashMap Text Text))

A collection of key-value pairs.