Amazonka.MediaLive.Types.M2tsSettings

Description

Documentation

data M2tsSettings

M2ts Settings

See: newM2tsSettings smart constructor.

Constructors

M2tsSettings'

Fields

pmtPid :: Maybe Text

Packet Identifier (PID) for the Program Map Table (PMT) in the transport stream. Can be entered as a decimal or hexadecimal value. Valid values are 32 (or 0x20)..8182 (or 0x1ff6).

etvSignalPid :: Maybe Text

Packet Identifier (PID) for input source ETV Signal data to this output. Can be entered as a decimal or hexadecimal value. Valid values are 32 (or 0x20)..8182 (or 0x1ff6).

videoPid :: Maybe Text

Packet Identifier (PID) of the elementary video stream in the transport stream. Can be entered as a decimal or hexadecimal value. Valid values are 32 (or 0x20)..8182 (or 0x1ff6).

nielsenId3Behavior :: Maybe M2tsNielsenId3Behavior

If set to passthrough, Nielsen inaudible tones for media tracking will be detected in the input audio and an equivalent ID3 tag will be inserted in the output.

bufferModel :: Maybe M2tsBufferModel

Controls the timing accuracy for output network traffic. Leave as MULTIPLEX to ensure accurate network packet timing. Or set to NONE, which might result in lower latency but will result in more variability in output network packet timing. This variability might cause interruptions, jitter, or bursty behavior in your playback or receiving devices.

scte35Pid :: Maybe Text

Packet Identifier (PID) of the SCTE-35 stream in the transport stream. Can be entered as a decimal or hexadecimal value. Valid values are 32 (or 0x20)..8182 (or 0x1ff6).

transportStreamId :: Maybe Natural

The value of the transport stream ID field in the Program Map Table.

programNum :: Maybe Natural

The value of the program number field in the Program Map Table.

fragmentTime :: Maybe Double

The length in seconds of each fragment. Only used with EBP markers.

timedMetadataBehavior :: Maybe M2tsTimedMetadataBehavior

When set to passthrough, timed metadata will be passed through from input to output.

ccDescriptor :: Maybe M2tsCcDescriptor

When set to enabled, generates captionServiceDescriptor in PMT.

pmtInterval :: Maybe Natural

The number of milliseconds between instances of this table in the output transport stream. Valid values are 0, 10..1000.

dvbSdtSettings :: Maybe DvbSdtSettings

Inserts DVB Service Description Table (SDT) at the specified table repetition interval.

ecmPid :: Maybe Text

This field is unused and deprecated.

nullPacketBitrate :: Maybe Double

Value in bits per second of extra null packets to insert into the transport stream. This can be used if a downstream encryption system requires periodic null packets.

audioBufferModel :: Maybe M2tsAudioBufferModel

When set to dvb, uses DVB buffer model for Dolby Digital audio. When set to atsc, the ATSC model is used.

timedMetadataPid :: Maybe Text

Packet Identifier (PID) of the timed metadata stream in the transport stream. Can be entered as a decimal or hexadecimal value. Valid values are 32 (or 0x20)..8182 (or 0x1ff6).

klv :: Maybe M2tsKlv

If set to passthrough, passes any KLV data from the input source to this output.

audioFramesPerPes :: Maybe Natural

The number of audio frames to insert for each PES packet.

pcrPeriod :: Maybe Natural

Maximum time in milliseconds between Program Clock Reference (PCRs) inserted into the transport stream.

pcrPid :: Maybe Text

Packet Identifier (PID) of the Program Clock Reference (PCR) in the transport stream. When no value is given, the encoder will assign the same value as the Video PID. Can be entered as a decimal or hexadecimal value. Valid values are 32 (or 0x20)..8182 (or 0x1ff6).

segmentationMarkers :: Maybe M2tsSegmentationMarkers

Inserts segmentation markers at each segmentationTime period. raiSegstart sets the Random Access Indicator bit in the adaptation field. raiAdapt sets the RAI bit and adds the current timecode in the private data bytes. psiSegstart inserts PAT and PMT tables at the start of segments. ebp adds Encoder Boundary Point information to the adaptation field as per OpenCable specification OC-SP-EBP-I01-130118. ebpLegacy adds Encoder Boundary Point information to the adaptation field using a legacy proprietary format.

aribCaptionsPidControl :: Maybe M2tsAribCaptionsPidControl

If set to auto, pid number used for ARIB Captions will be auto-selected from unused pids. If set to useConfigured, ARIB Captions will be on the configured pid number.

klvDataPids :: Maybe Text

Packet Identifier (PID) for input source KLV data to this output. Multiple values are accepted, and can be entered in ranges and/or by comma separation. Can be entered as decimal or hexadecimal values. Each PID specified must be in the range of 32 (or 0x20)..8182 (or 0x1ff6).

ebpLookaheadMs :: Maybe Natural

When set, enforces that Encoder Boundary Points do not come within the specified time interval of each other by looking ahead at input video. If another EBP is going to come in within the specified time interval, the current EBP is not emitted, and the segment is "stretched" to the next marker. The lookahead value does not add latency to the system. The Live Event must be configured elsewhere to create sufficient latency to make the lookahead accurate.

dvbSubPids :: Maybe Text

Packet Identifier (PID) for input source DVB Subtitle data to this output. Multiple values are accepted, and can be entered in ranges and/or by comma separation. Can be entered as decimal or hexadecimal values. Each PID specified must be in the range of 32 (or 0x20)..8182 (or 0x1ff6).

scte27Pids :: Maybe Text

Packet Identifier (PID) for input source SCTE-27 data to this output. Multiple values are accepted, and can be entered in ranges and/or by comma separation. Can be entered as decimal or hexadecimal values. Each PID specified must be in the range of 32 (or 0x20)..8182 (or 0x1ff6).

patInterval :: Maybe Natural

The number of milliseconds between instances of this table in the output transport stream. Valid values are 0, 10..1000.

audioStreamType :: Maybe M2tsAudioStreamType

When set to atsc, uses stream type = 0x81 for AC3 and stream type = 0x87 for EAC3. When set to dvb, uses stream type = 0x06.

esRateInPes :: Maybe M2tsEsRateInPes

Include or exclude the ES Rate field in the PES header.

etvPlatformPid :: Maybe Text

Packet Identifier (PID) for input source ETV Platform data to this output. Can be entered as a decimal or hexadecimal value. Valid values are 32 (or 0x20)..8182 (or 0x1ff6).

bitrate :: Maybe Natural

The output bitrate of the transport stream in bits per second. Setting to 0 lets the muxer automatically determine the appropriate bitrate.

scte35Control :: Maybe M2tsScte35Control

Optionally pass SCTE-35 signals from the input source to this output.

audioPids :: Maybe Text

Packet Identifier (PID) of the elementary audio stream(s) in the transport stream. Multiple values are accepted, and can be entered in ranges and/or by comma separation. Can be entered as decimal or hexadecimal values. Each PID specified must be in the range of 32 (or 0x20)..8182 (or 0x1ff6).

dvbTeletextPid :: Maybe Text

Packet Identifier (PID) for input source DVB Teletext data to this output. Can be entered as a decimal or hexadecimal value. Valid values are 32 (or 0x20)..8182 (or 0x1ff6).

ebif :: Maybe M2tsEbifControl

If set to passthrough, passes any EBIF data from the input source to this output.

arib :: Maybe M2tsArib

When set to enabled, uses ARIB-compliant field muxing and removes video descriptor.

aribCaptionsPid :: Maybe Text

Packet Identifier (PID) for ARIB Captions in the transport stream. Can be entered as a decimal or hexadecimal value. Valid values are 32 (or 0x20)..8182 (or 0x1ff6).

absentInputAudioBehavior :: Maybe M2tsAbsentInputAudioBehavior

When set to drop, output audio streams will be removed from the program if the selected input audio stream is removed from the input. This allows the output audio configuration to dynamically change based on input configuration. If this is set to encodeSilence, all output audio streams will output encoded silence when not connected to an active input stream.

segmentationTime :: Maybe Double

The length in seconds of each segment. Required unless markers is set to _none_.

ebpAudioInterval :: Maybe M2tsAudioInterval

When videoAndFixedIntervals is selected, audio EBP markers will be added to partitions 3 and 4. The interval between these additional markers will be fixed, and will be slightly shorter than the video EBP marker interval. Only available when EBP Cablelabs segmentation markers are selected. Partitions 1 and 2 will always follow the video interval.

dvbNitSettings :: Maybe DvbNitSettings

Inserts DVB Network Information Table (NIT) at the specified table repetition interval.

pcrControl :: Maybe M2tsPcrControl

When set to pcrEveryPesPacket, a Program Clock Reference value is inserted for every Packetized Elementary Stream (PES) header. This parameter is effective only when the PCR PID is the same as the video or audio elementary stream.

ebpPlacement :: Maybe M2tsEbpPlacement

Controls placement of EBP on Audio PIDs. If set to videoAndAudioPids, EBP markers will be placed on the video PID and all audio PIDs. If set to videoPid, EBP markers will be placed on only the video PID.

rateMode :: Maybe M2tsRateMode

When vbr, does not insert null packets into transport stream to fill specified bitrate. The bitrate setting acts as the maximum bitrate when vbr is set.

segmentationStyle :: Maybe M2tsSegmentationStyle

The segmentation style parameter controls how segmentation markers are inserted into the transport stream. With avails, it is possible that segments may be truncated, which can influence where future segmentation markers are inserted. When a segmentation style of "resetCadence" is selected and a segment is truncated due to an avail, we will reset the segmentation cadence. This means the subsequent segment will have a duration of $segmentationTime seconds. When a segmentation style of "maintainCadence" is selected and a segment is truncated due to an avail, we will not reset the segmentation cadence. This means the subsequent segment will likely be truncated as well. However, all segments after that will have a duration of $segmentationTime seconds. Note that EBP lookahead is a slight exception to this rule.

dvbTdtSettings :: Maybe DvbTdtSettings

Inserts DVB Time and Date Table (TDT) at the specified table repetition interval.

newM2tsSettings :: M2tsSettings

Create a value of M2tsSettings with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:pmtPid:M2tsSettings', m2tsSettings_pmtPid - Packet Identifier (PID) for the Program Map Table (PMT) in the transport stream. Can be entered as a decimal or hexadecimal value. Valid values are 32 (or 0x20)..8182 (or 0x1ff6).

$sel:etvSignalPid:M2tsSettings', m2tsSettings_etvSignalPid - Packet Identifier (PID) for input source ETV Signal data to this output. Can be entered as a decimal or hexadecimal value. Valid values are 32 (or 0x20)..8182 (or 0x1ff6).

$sel:videoPid:M2tsSettings', m2tsSettings_videoPid - Packet Identifier (PID) of the elementary video stream in the transport stream. Can be entered as a decimal or hexadecimal value. Valid values are 32 (or 0x20)..8182 (or 0x1ff6).

$sel:nielsenId3Behavior:M2tsSettings', m2tsSettings_nielsenId3Behavior - If set to passthrough, Nielsen inaudible tones for media tracking will be detected in the input audio and an equivalent ID3 tag will be inserted in the output.

$sel:bufferModel:M2tsSettings', m2tsSettings_bufferModel - Controls the timing accuracy for output network traffic. Leave as MULTIPLEX to ensure accurate network packet timing. Or set to NONE, which might result in lower latency but will result in more variability in output network packet timing. This variability might cause interruptions, jitter, or bursty behavior in your playback or receiving devices.

$sel:scte35Pid:M2tsSettings', m2tsSettings_scte35Pid - Packet Identifier (PID) of the SCTE-35 stream in the transport stream. Can be entered as a decimal or hexadecimal value. Valid values are 32 (or 0x20)..8182 (or 0x1ff6).

$sel:transportStreamId:M2tsSettings', m2tsSettings_transportStreamId - The value of the transport stream ID field in the Program Map Table.

$sel:programNum:M2tsSettings', m2tsSettings_programNum - The value of the program number field in the Program Map Table.

$sel:fragmentTime:M2tsSettings', m2tsSettings_fragmentTime - The length in seconds of each fragment. Only used with EBP markers.

$sel:timedMetadataBehavior:M2tsSettings', m2tsSettings_timedMetadataBehavior - When set to passthrough, timed metadata will be passed through from input to output.

$sel:ccDescriptor:M2tsSettings', m2tsSettings_ccDescriptor - When set to enabled, generates captionServiceDescriptor in PMT.

$sel:pmtInterval:M2tsSettings', m2tsSettings_pmtInterval - The number of milliseconds between instances of this table in the output transport stream. Valid values are 0, 10..1000.

$sel:dvbSdtSettings:M2tsSettings', m2tsSettings_dvbSdtSettings - Inserts DVB Service Description Table (SDT) at the specified table repetition interval.

$sel:ecmPid:M2tsSettings', m2tsSettings_ecmPid - This field is unused and deprecated.

$sel:nullPacketBitrate:M2tsSettings', m2tsSettings_nullPacketBitrate - Value in bits per second of extra null packets to insert into the transport stream. This can be used if a downstream encryption system requires periodic null packets.

$sel:audioBufferModel:M2tsSettings', m2tsSettings_audioBufferModel - When set to dvb, uses DVB buffer model for Dolby Digital audio. When set to atsc, the ATSC model is used.

$sel:timedMetadataPid:M2tsSettings', m2tsSettings_timedMetadataPid - Packet Identifier (PID) of the timed metadata stream in the transport stream. Can be entered as a decimal or hexadecimal value. Valid values are 32 (or 0x20)..8182 (or 0x1ff6).

$sel:klv:M2tsSettings', m2tsSettings_klv - If set to passthrough, passes any KLV data from the input source to this output.

$sel:audioFramesPerPes:M2tsSettings', m2tsSettings_audioFramesPerPes - The number of audio frames to insert for each PES packet.

$sel:pcrPeriod:M2tsSettings', m2tsSettings_pcrPeriod - Maximum time in milliseconds between Program Clock Reference (PCRs) inserted into the transport stream.

$sel:pcrPid:M2tsSettings', m2tsSettings_pcrPid - Packet Identifier (PID) of the Program Clock Reference (PCR) in the transport stream. When no value is given, the encoder will assign the same value as the Video PID. Can be entered as a decimal or hexadecimal value. Valid values are 32 (or 0x20)..8182 (or 0x1ff6).

$sel:segmentationMarkers:M2tsSettings', m2tsSettings_segmentationMarkers - Inserts segmentation markers at each segmentationTime period. raiSegstart sets the Random Access Indicator bit in the adaptation field. raiAdapt sets the RAI bit and adds the current timecode in the private data bytes. psiSegstart inserts PAT and PMT tables at the start of segments. ebp adds Encoder Boundary Point information to the adaptation field as per OpenCable specification OC-SP-EBP-I01-130118. ebpLegacy adds Encoder Boundary Point information to the adaptation field using a legacy proprietary format.

$sel:aribCaptionsPidControl:M2tsSettings', m2tsSettings_aribCaptionsPidControl - If set to auto, pid number used for ARIB Captions will be auto-selected from unused pids. If set to useConfigured, ARIB Captions will be on the configured pid number.

$sel:klvDataPids:M2tsSettings', m2tsSettings_klvDataPids - Packet Identifier (PID) for input source KLV data to this output. Multiple values are accepted, and can be entered in ranges and/or by comma separation. Can be entered as decimal or hexadecimal values. Each PID specified must be in the range of 32 (or 0x20)..8182 (or 0x1ff6).

$sel:ebpLookaheadMs:M2tsSettings', m2tsSettings_ebpLookaheadMs - When set, enforces that Encoder Boundary Points do not come within the specified time interval of each other by looking ahead at input video. If another EBP is going to come in within the specified time interval, the current EBP is not emitted, and the segment is "stretched" to the next marker. The lookahead value does not add latency to the system. The Live Event must be configured elsewhere to create sufficient latency to make the lookahead accurate.

$sel:dvbSubPids:M2tsSettings', m2tsSettings_dvbSubPids - Packet Identifier (PID) for input source DVB Subtitle data to this output. Multiple values are accepted, and can be entered in ranges and/or by comma separation. Can be entered as decimal or hexadecimal values. Each PID specified must be in the range of 32 (or 0x20)..8182 (or 0x1ff6).

$sel:scte27Pids:M2tsSettings', m2tsSettings_scte27Pids - Packet Identifier (PID) for input source SCTE-27 data to this output. Multiple values are accepted, and can be entered in ranges and/or by comma separation. Can be entered as decimal or hexadecimal values. Each PID specified must be in the range of 32 (or 0x20)..8182 (or 0x1ff6).

$sel:patInterval:M2tsSettings', m2tsSettings_patInterval - The number of milliseconds between instances of this table in the output transport stream. Valid values are 0, 10..1000.

$sel:audioStreamType:M2tsSettings', m2tsSettings_audioStreamType - When set to atsc, uses stream type = 0x81 for AC3 and stream type = 0x87 for EAC3. When set to dvb, uses stream type = 0x06.

$sel:esRateInPes:M2tsSettings', m2tsSettings_esRateInPes - Include or exclude the ES Rate field in the PES header.

$sel:etvPlatformPid:M2tsSettings', m2tsSettings_etvPlatformPid - Packet Identifier (PID) for input source ETV Platform data to this output. Can be entered as a decimal or hexadecimal value. Valid values are 32 (or 0x20)..8182 (or 0x1ff6).

$sel:bitrate:M2tsSettings', m2tsSettings_bitrate - The output bitrate of the transport stream in bits per second. Setting to 0 lets the muxer automatically determine the appropriate bitrate.

$sel:scte35Control:M2tsSettings', m2tsSettings_scte35Control - Optionally pass SCTE-35 signals from the input source to this output.

$sel:audioPids:M2tsSettings', m2tsSettings_audioPids - Packet Identifier (PID) of the elementary audio stream(s) in the transport stream. Multiple values are accepted, and can be entered in ranges and/or by comma separation. Can be entered as decimal or hexadecimal values. Each PID specified must be in the range of 32 (or 0x20)..8182 (or 0x1ff6).

$sel:dvbTeletextPid:M2tsSettings', m2tsSettings_dvbTeletextPid - Packet Identifier (PID) for input source DVB Teletext data to this output. Can be entered as a decimal or hexadecimal value. Valid values are 32 (or 0x20)..8182 (or 0x1ff6).

$sel:ebif:M2tsSettings', m2tsSettings_ebif - If set to passthrough, passes any EBIF data from the input source to this output.

$sel:arib:M2tsSettings', m2tsSettings_arib - When set to enabled, uses ARIB-compliant field muxing and removes video descriptor.

$sel:aribCaptionsPid:M2tsSettings', m2tsSettings_aribCaptionsPid - Packet Identifier (PID) for ARIB Captions in the transport stream. Can be entered as a decimal or hexadecimal value. Valid values are 32 (or 0x20)..8182 (or 0x1ff6).

$sel:absentInputAudioBehavior:M2tsSettings', m2tsSettings_absentInputAudioBehavior - When set to drop, output audio streams will be removed from the program if the selected input audio stream is removed from the input. This allows the output audio configuration to dynamically change based on input configuration. If this is set to encodeSilence, all output audio streams will output encoded silence when not connected to an active input stream.

$sel:segmentationTime:M2tsSettings', m2tsSettings_segmentationTime - The length in seconds of each segment. Required unless markers is set to _none_.

$sel:ebpAudioInterval:M2tsSettings', m2tsSettings_ebpAudioInterval - When videoAndFixedIntervals is selected, audio EBP markers will be added to partitions 3 and 4. The interval between these additional markers will be fixed, and will be slightly shorter than the video EBP marker interval. Only available when EBP Cablelabs segmentation markers are selected. Partitions 1 and 2 will always follow the video interval.

$sel:dvbNitSettings:M2tsSettings', m2tsSettings_dvbNitSettings - Inserts DVB Network Information Table (NIT) at the specified table repetition interval.

$sel:pcrControl:M2tsSettings', m2tsSettings_pcrControl - When set to pcrEveryPesPacket, a Program Clock Reference value is inserted for every Packetized Elementary Stream (PES) header. This parameter is effective only when the PCR PID is the same as the video or audio elementary stream.

$sel:ebpPlacement:M2tsSettings', m2tsSettings_ebpPlacement - Controls placement of EBP on Audio PIDs. If set to videoAndAudioPids, EBP markers will be placed on the video PID and all audio PIDs. If set to videoPid, EBP markers will be placed on only the video PID.

$sel:rateMode:M2tsSettings', m2tsSettings_rateMode - When vbr, does not insert null packets into transport stream to fill specified bitrate. The bitrate setting acts as the maximum bitrate when vbr is set.

$sel:segmentationStyle:M2tsSettings', m2tsSettings_segmentationStyle - The segmentation style parameter controls how segmentation markers are inserted into the transport stream. With avails, it is possible that segments may be truncated, which can influence where future segmentation markers are inserted. When a segmentation style of "resetCadence" is selected and a segment is truncated due to an avail, we will reset the segmentation cadence. This means the subsequent segment will have a duration of $segmentationTime seconds. When a segmentation style of "maintainCadence" is selected and a segment is truncated due to an avail, we will not reset the segmentation cadence. This means the subsequent segment will likely be truncated as well. However, all segments after that will have a duration of $segmentationTime seconds. Note that EBP lookahead is a slight exception to this rule.

$sel:dvbTdtSettings:M2tsSettings', m2tsSettings_dvbTdtSettings - Inserts DVB Time and Date Table (TDT) at the specified table repetition interval.

m2tsSettings_pmtPid :: Lens' M2tsSettings (Maybe Text)

Packet Identifier (PID) for the Program Map Table (PMT) in the transport stream. Can be entered as a decimal or hexadecimal value. Valid values are 32 (or 0x20)..8182 (or 0x1ff6).

m2tsSettings_etvSignalPid :: Lens' M2tsSettings (Maybe Text)

Packet Identifier (PID) for input source ETV Signal data to this output. Can be entered as a decimal or hexadecimal value. Valid values are 32 (or 0x20)..8182 (or 0x1ff6).

m2tsSettings_videoPid :: Lens' M2tsSettings (Maybe Text)

Packet Identifier (PID) of the elementary video stream in the transport stream. Can be entered as a decimal or hexadecimal value. Valid values are 32 (or 0x20)..8182 (or 0x1ff6).

m2tsSettings_nielsenId3Behavior :: Lens' M2tsSettings (Maybe M2tsNielsenId3Behavior)

If set to passthrough, Nielsen inaudible tones for media tracking will be detected in the input audio and an equivalent ID3 tag will be inserted in the output.

m2tsSettings_bufferModel :: Lens' M2tsSettings (Maybe M2tsBufferModel)

Controls the timing accuracy for output network traffic. Leave as MULTIPLEX to ensure accurate network packet timing. Or set to NONE, which might result in lower latency but will result in more variability in output network packet timing. This variability might cause interruptions, jitter, or bursty behavior in your playback or receiving devices.

m2tsSettings_scte35Pid :: Lens' M2tsSettings (Maybe Text)

Packet Identifier (PID) of the SCTE-35 stream in the transport stream. Can be entered as a decimal or hexadecimal value. Valid values are 32 (or 0x20)..8182 (or 0x1ff6).

m2tsSettings_transportStreamId :: Lens' M2tsSettings (Maybe Natural)

The value of the transport stream ID field in the Program Map Table.

m2tsSettings_programNum :: Lens' M2tsSettings (Maybe Natural)

The value of the program number field in the Program Map Table.

m2tsSettings_fragmentTime :: Lens' M2tsSettings (Maybe Double)

The length in seconds of each fragment. Only used with EBP markers.

m2tsSettings_timedMetadataBehavior :: Lens' M2tsSettings (Maybe M2tsTimedMetadataBehavior)

When set to passthrough, timed metadata will be passed through from input to output.

m2tsSettings_ccDescriptor :: Lens' M2tsSettings (Maybe M2tsCcDescriptor)

When set to enabled, generates captionServiceDescriptor in PMT.

m2tsSettings_pmtInterval :: Lens' M2tsSettings (Maybe Natural)

The number of milliseconds between instances of this table in the output transport stream. Valid values are 0, 10..1000.

m2tsSettings_dvbSdtSettings :: Lens' M2tsSettings (Maybe DvbSdtSettings)

Inserts DVB Service Description Table (SDT) at the specified table repetition interval.

m2tsSettings_ecmPid :: Lens' M2tsSettings (Maybe Text)

This field is unused and deprecated.

m2tsSettings_nullPacketBitrate :: Lens' M2tsSettings (Maybe Double)

Value in bits per second of extra null packets to insert into the transport stream. This can be used if a downstream encryption system requires periodic null packets.

m2tsSettings_audioBufferModel :: Lens' M2tsSettings (Maybe M2tsAudioBufferModel)

When set to dvb, uses DVB buffer model for Dolby Digital audio. When set to atsc, the ATSC model is used.

m2tsSettings_timedMetadataPid :: Lens' M2tsSettings (Maybe Text)

Packet Identifier (PID) of the timed metadata stream in the transport stream. Can be entered as a decimal or hexadecimal value. Valid values are 32 (or 0x20)..8182 (or 0x1ff6).

m2tsSettings_klv :: Lens' M2tsSettings (Maybe M2tsKlv)

If set to passthrough, passes any KLV data from the input source to this output.

m2tsSettings_audioFramesPerPes :: Lens' M2tsSettings (Maybe Natural)

The number of audio frames to insert for each PES packet.

m2tsSettings_pcrPeriod :: Lens' M2tsSettings (Maybe Natural)

Maximum time in milliseconds between Program Clock Reference (PCRs) inserted into the transport stream.

m2tsSettings_pcrPid :: Lens' M2tsSettings (Maybe Text)

Packet Identifier (PID) of the Program Clock Reference (PCR) in the transport stream. When no value is given, the encoder will assign the same value as the Video PID. Can be entered as a decimal or hexadecimal value. Valid values are 32 (or 0x20)..8182 (or 0x1ff6).

m2tsSettings_segmentationMarkers :: Lens' M2tsSettings (Maybe M2tsSegmentationMarkers)

Inserts segmentation markers at each segmentationTime period. raiSegstart sets the Random Access Indicator bit in the adaptation field. raiAdapt sets the RAI bit and adds the current timecode in the private data bytes. psiSegstart inserts PAT and PMT tables at the start of segments. ebp adds Encoder Boundary Point information to the adaptation field as per OpenCable specification OC-SP-EBP-I01-130118. ebpLegacy adds Encoder Boundary Point information to the adaptation field using a legacy proprietary format.

m2tsSettings_aribCaptionsPidControl :: Lens' M2tsSettings (Maybe M2tsAribCaptionsPidControl)

If set to auto, pid number used for ARIB Captions will be auto-selected from unused pids. If set to useConfigured, ARIB Captions will be on the configured pid number.

m2tsSettings_klvDataPids :: Lens' M2tsSettings (Maybe Text)

Packet Identifier (PID) for input source KLV data to this output. Multiple values are accepted, and can be entered in ranges and/or by comma separation. Can be entered as decimal or hexadecimal values. Each PID specified must be in the range of 32 (or 0x20)..8182 (or 0x1ff6).

m2tsSettings_ebpLookaheadMs :: Lens' M2tsSettings (Maybe Natural)

When set, enforces that Encoder Boundary Points do not come within the specified time interval of each other by looking ahead at input video. If another EBP is going to come in within the specified time interval, the current EBP is not emitted, and the segment is "stretched" to the next marker. The lookahead value does not add latency to the system. The Live Event must be configured elsewhere to create sufficient latency to make the lookahead accurate.

m2tsSettings_dvbSubPids :: Lens' M2tsSettings (Maybe Text)

Packet Identifier (PID) for input source DVB Subtitle data to this output. Multiple values are accepted, and can be entered in ranges and/or by comma separation. Can be entered as decimal or hexadecimal values. Each PID specified must be in the range of 32 (or 0x20)..8182 (or 0x1ff6).

m2tsSettings_scte27Pids :: Lens' M2tsSettings (Maybe Text)

Packet Identifier (PID) for input source SCTE-27 data to this output. Multiple values are accepted, and can be entered in ranges and/or by comma separation. Can be entered as decimal or hexadecimal values. Each PID specified must be in the range of 32 (or 0x20)..8182 (or 0x1ff6).

m2tsSettings_patInterval :: Lens' M2tsSettings (Maybe Natural)

The number of milliseconds between instances of this table in the output transport stream. Valid values are 0, 10..1000.

m2tsSettings_audioStreamType :: Lens' M2tsSettings (Maybe M2tsAudioStreamType)

When set to atsc, uses stream type = 0x81 for AC3 and stream type = 0x87 for EAC3. When set to dvb, uses stream type = 0x06.

m2tsSettings_esRateInPes :: Lens' M2tsSettings (Maybe M2tsEsRateInPes)

Include or exclude the ES Rate field in the PES header.

m2tsSettings_etvPlatformPid :: Lens' M2tsSettings (Maybe Text)

Packet Identifier (PID) for input source ETV Platform data to this output. Can be entered as a decimal or hexadecimal value. Valid values are 32 (or 0x20)..8182 (or 0x1ff6).

m2tsSettings_bitrate :: Lens' M2tsSettings (Maybe Natural)

The output bitrate of the transport stream in bits per second. Setting to 0 lets the muxer automatically determine the appropriate bitrate.

m2tsSettings_scte35Control :: Lens' M2tsSettings (Maybe M2tsScte35Control)

Optionally pass SCTE-35 signals from the input source to this output.

m2tsSettings_audioPids :: Lens' M2tsSettings (Maybe Text)

Packet Identifier (PID) of the elementary audio stream(s) in the transport stream. Multiple values are accepted, and can be entered in ranges and/or by comma separation. Can be entered as decimal or hexadecimal values. Each PID specified must be in the range of 32 (or 0x20)..8182 (or 0x1ff6).

m2tsSettings_dvbTeletextPid :: Lens' M2tsSettings (Maybe Text)

Packet Identifier (PID) for input source DVB Teletext data to this output. Can be entered as a decimal or hexadecimal value. Valid values are 32 (or 0x20)..8182 (or 0x1ff6).

m2tsSettings_ebif :: Lens' M2tsSettings (Maybe M2tsEbifControl)

If set to passthrough, passes any EBIF data from the input source to this output.

m2tsSettings_arib :: Lens' M2tsSettings (Maybe M2tsArib)

When set to enabled, uses ARIB-compliant field muxing and removes video descriptor.

m2tsSettings_aribCaptionsPid :: Lens' M2tsSettings (Maybe Text)

Packet Identifier (PID) for ARIB Captions in the transport stream. Can be entered as a decimal or hexadecimal value. Valid values are 32 (or 0x20)..8182 (or 0x1ff6).

m2tsSettings_absentInputAudioBehavior :: Lens' M2tsSettings (Maybe M2tsAbsentInputAudioBehavior)

When set to drop, output audio streams will be removed from the program if the selected input audio stream is removed from the input. This allows the output audio configuration to dynamically change based on input configuration. If this is set to encodeSilence, all output audio streams will output encoded silence when not connected to an active input stream.

m2tsSettings_segmentationTime :: Lens' M2tsSettings (Maybe Double)

The length in seconds of each segment. Required unless markers is set to _none_.

m2tsSettings_ebpAudioInterval :: Lens' M2tsSettings (Maybe M2tsAudioInterval)

When videoAndFixedIntervals is selected, audio EBP markers will be added to partitions 3 and 4. The interval between these additional markers will be fixed, and will be slightly shorter than the video EBP marker interval. Only available when EBP Cablelabs segmentation markers are selected. Partitions 1 and 2 will always follow the video interval.

m2tsSettings_dvbNitSettings :: Lens' M2tsSettings (Maybe DvbNitSettings)

Inserts DVB Network Information Table (NIT) at the specified table repetition interval.

m2tsSettings_pcrControl :: Lens' M2tsSettings (Maybe M2tsPcrControl)

When set to pcrEveryPesPacket, a Program Clock Reference value is inserted for every Packetized Elementary Stream (PES) header. This parameter is effective only when the PCR PID is the same as the video or audio elementary stream.

m2tsSettings_ebpPlacement :: Lens' M2tsSettings (Maybe M2tsEbpPlacement)

Controls placement of EBP on Audio PIDs. If set to videoAndAudioPids, EBP markers will be placed on the video PID and all audio PIDs. If set to videoPid, EBP markers will be placed on only the video PID.

m2tsSettings_rateMode :: Lens' M2tsSettings (Maybe M2tsRateMode)

When vbr, does not insert null packets into transport stream to fill specified bitrate. The bitrate setting acts as the maximum bitrate when vbr is set.

m2tsSettings_segmentationStyle :: Lens' M2tsSettings (Maybe M2tsSegmentationStyle)

The segmentation style parameter controls how segmentation markers are inserted into the transport stream. With avails, it is possible that segments may be truncated, which can influence where future segmentation markers are inserted. When a segmentation style of "resetCadence" is selected and a segment is truncated due to an avail, we will reset the segmentation cadence. This means the subsequent segment will have a duration of $segmentationTime seconds. When a segmentation style of "maintainCadence" is selected and a segment is truncated due to an avail, we will not reset the segmentation cadence. This means the subsequent segment will likely be truncated as well. However, all segments after that will have a duration of $segmentationTime seconds. Note that EBP lookahead is a slight exception to this rule.

m2tsSettings_dvbTdtSettings :: Lens' M2tsSettings (Maybe DvbTdtSettings)

Inserts DVB Time and Date Table (TDT) at the specified table repetition interval.