Amazonka.MediaLive.Types.M3u8Settings

Description

Documentation

data M3u8Settings

Settings information for the .m3u8 container

See: newM3u8Settings smart constructor.

Constructors

M3u8Settings'

Fields

pmtPid :: Maybe Text

Packet Identifier (PID) for the Program Map Table (PMT) in the transport stream. Can be entered as a decimal or hexadecimal value.

videoPid :: Maybe Text

Packet Identifier (PID) of the elementary video stream in the transport stream. Can be entered as a decimal or hexadecimal value.

nielsenId3Behavior :: Maybe M3u8NielsenId3Behavior

If set to passthrough, Nielsen inaudible tones for media tracking will be detected in the input audio and an equivalent ID3 tag will be inserted in the output.

scte35Pid :: Maybe Text

Packet Identifier (PID) of the SCTE-35 stream in the transport stream. Can be entered as a decimal or hexadecimal value.

transportStreamId :: Maybe Natural

The value of the transport stream ID field in the Program Map Table.

programNum :: Maybe Natural

The value of the program number field in the Program Map Table.

timedMetadataBehavior :: Maybe M3u8TimedMetadataBehavior

When set to passthrough, timed metadata is passed through from input to output.

pmtInterval :: Maybe Natural

The number of milliseconds between instances of this table in the output transport stream. A value of \"0\" writes out the PMT once per segment file.

ecmPid :: Maybe Text

This parameter is unused and deprecated.

timedMetadataPid :: Maybe Text

Packet Identifier (PID) of the timed metadata stream in the transport stream. Can be entered as a decimal or hexadecimal value. Valid values are 32 (or 0x20)..8182 (or 0x1ff6).

audioFramesPerPes :: Maybe Natural

The number of audio frames to insert for each PES packet.

pcrPeriod :: Maybe Natural

Maximum time in milliseconds between Program Clock References (PCRs) inserted into the transport stream.

pcrPid :: Maybe Text

Packet Identifier (PID) of the Program Clock Reference (PCR) in the transport stream. When no value is given, the encoder will assign the same value as the Video PID. Can be entered as a decimal or hexadecimal value.

patInterval :: Maybe Natural

The number of milliseconds between instances of this table in the output transport stream. A value of \"0\" writes out the PMT once per segment file.

audioPids :: Maybe Text

Packet Identifier (PID) of the elementary audio stream(s) in the transport stream. Multiple values are accepted, and can be entered in ranges and/or by comma separation. Can be entered as decimal or hexadecimal values.

scte35Behavior :: Maybe M3u8Scte35Behavior

If set to passthrough, passes any SCTE-35 signals from the input source to this output.

pcrControl :: Maybe M3u8PcrControl

When set to pcrEveryPesPacket, a Program Clock Reference value is inserted for every Packetized Elementary Stream (PES) header. This parameter is effective only when the PCR PID is the same as the video or audio elementary stream.

newM3u8Settings :: M3u8Settings

Create a value of M3u8Settings with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:pmtPid:M3u8Settings', m3u8Settings_pmtPid - Packet Identifier (PID) for the Program Map Table (PMT) in the transport stream. Can be entered as a decimal or hexadecimal value.

$sel:videoPid:M3u8Settings', m3u8Settings_videoPid - Packet Identifier (PID) of the elementary video stream in the transport stream. Can be entered as a decimal or hexadecimal value.

$sel:nielsenId3Behavior:M3u8Settings', m3u8Settings_nielsenId3Behavior - If set to passthrough, Nielsen inaudible tones for media tracking will be detected in the input audio and an equivalent ID3 tag will be inserted in the output.

$sel:scte35Pid:M3u8Settings', m3u8Settings_scte35Pid - Packet Identifier (PID) of the SCTE-35 stream in the transport stream. Can be entered as a decimal or hexadecimal value.

$sel:transportStreamId:M3u8Settings', m3u8Settings_transportStreamId - The value of the transport stream ID field in the Program Map Table.

$sel:programNum:M3u8Settings', m3u8Settings_programNum - The value of the program number field in the Program Map Table.

$sel:timedMetadataBehavior:M3u8Settings', m3u8Settings_timedMetadataBehavior - When set to passthrough, timed metadata is passed through from input to output.

$sel:pmtInterval:M3u8Settings', m3u8Settings_pmtInterval - The number of milliseconds between instances of this table in the output transport stream. A value of \"0\" writes out the PMT once per segment file.

$sel:ecmPid:M3u8Settings', m3u8Settings_ecmPid - This parameter is unused and deprecated.

$sel:timedMetadataPid:M3u8Settings', m3u8Settings_timedMetadataPid - Packet Identifier (PID) of the timed metadata stream in the transport stream. Can be entered as a decimal or hexadecimal value. Valid values are 32 (or 0x20)..8182 (or 0x1ff6).

$sel:audioFramesPerPes:M3u8Settings', m3u8Settings_audioFramesPerPes - The number of audio frames to insert for each PES packet.

$sel:pcrPeriod:M3u8Settings', m3u8Settings_pcrPeriod - Maximum time in milliseconds between Program Clock References (PCRs) inserted into the transport stream.

$sel:pcrPid:M3u8Settings', m3u8Settings_pcrPid - Packet Identifier (PID) of the Program Clock Reference (PCR) in the transport stream. When no value is given, the encoder will assign the same value as the Video PID. Can be entered as a decimal or hexadecimal value.

$sel:patInterval:M3u8Settings', m3u8Settings_patInterval - The number of milliseconds between instances of this table in the output transport stream. A value of \"0\" writes out the PMT once per segment file.

$sel:audioPids:M3u8Settings', m3u8Settings_audioPids - Packet Identifier (PID) of the elementary audio stream(s) in the transport stream. Multiple values are accepted, and can be entered in ranges and/or by comma separation. Can be entered as decimal or hexadecimal values.

$sel:scte35Behavior:M3u8Settings', m3u8Settings_scte35Behavior - If set to passthrough, passes any SCTE-35 signals from the input source to this output.

$sel:pcrControl:M3u8Settings', m3u8Settings_pcrControl - When set to pcrEveryPesPacket, a Program Clock Reference value is inserted for every Packetized Elementary Stream (PES) header. This parameter is effective only when the PCR PID is the same as the video or audio elementary stream.

m3u8Settings_pmtPid :: Lens' M3u8Settings (Maybe Text)

Packet Identifier (PID) for the Program Map Table (PMT) in the transport stream. Can be entered as a decimal or hexadecimal value.

m3u8Settings_videoPid :: Lens' M3u8Settings (Maybe Text)

Packet Identifier (PID) of the elementary video stream in the transport stream. Can be entered as a decimal or hexadecimal value.

m3u8Settings_nielsenId3Behavior :: Lens' M3u8Settings (Maybe M3u8NielsenId3Behavior)

If set to passthrough, Nielsen inaudible tones for media tracking will be detected in the input audio and an equivalent ID3 tag will be inserted in the output.

m3u8Settings_scte35Pid :: Lens' M3u8Settings (Maybe Text)

Packet Identifier (PID) of the SCTE-35 stream in the transport stream. Can be entered as a decimal or hexadecimal value.

m3u8Settings_transportStreamId :: Lens' M3u8Settings (Maybe Natural)

The value of the transport stream ID field in the Program Map Table.

m3u8Settings_programNum :: Lens' M3u8Settings (Maybe Natural)

The value of the program number field in the Program Map Table.

m3u8Settings_timedMetadataBehavior :: Lens' M3u8Settings (Maybe M3u8TimedMetadataBehavior)

When set to passthrough, timed metadata is passed through from input to output.

m3u8Settings_pmtInterval :: Lens' M3u8Settings (Maybe Natural)

The number of milliseconds between instances of this table in the output transport stream. A value of \"0\" writes out the PMT once per segment file.

m3u8Settings_ecmPid :: Lens' M3u8Settings (Maybe Text)

This parameter is unused and deprecated.

m3u8Settings_timedMetadataPid :: Lens' M3u8Settings (Maybe Text)

Packet Identifier (PID) of the timed metadata stream in the transport stream. Can be entered as a decimal or hexadecimal value. Valid values are 32 (or 0x20)..8182 (or 0x1ff6).

m3u8Settings_audioFramesPerPes :: Lens' M3u8Settings (Maybe Natural)

The number of audio frames to insert for each PES packet.

m3u8Settings_pcrPeriod :: Lens' M3u8Settings (Maybe Natural)

Maximum time in milliseconds between Program Clock References (PCRs) inserted into the transport stream.

m3u8Settings_pcrPid :: Lens' M3u8Settings (Maybe Text)

Packet Identifier (PID) of the Program Clock Reference (PCR) in the transport stream. When no value is given, the encoder will assign the same value as the Video PID. Can be entered as a decimal or hexadecimal value.

m3u8Settings_patInterval :: Lens' M3u8Settings (Maybe Natural)

The number of milliseconds between instances of this table in the output transport stream. A value of \"0\" writes out the PMT once per segment file.

m3u8Settings_audioPids :: Lens' M3u8Settings (Maybe Text)

Packet Identifier (PID) of the elementary audio stream(s) in the transport stream. Multiple values are accepted, and can be entered in ranges and/or by comma separation. Can be entered as decimal or hexadecimal values.

m3u8Settings_scte35Behavior :: Lens' M3u8Settings (Maybe M3u8Scte35Behavior)

If set to passthrough, passes any SCTE-35 signals from the input source to this output.

m3u8Settings_pcrControl :: Lens' M3u8Settings (Maybe M3u8PcrControl)

When set to pcrEveryPesPacket, a Program Clock Reference value is inserted for every Packetized Elementary Stream (PES) header. This parameter is effective only when the PCR PID is the same as the video or audio elementary stream.