Amazonka.MediaLive.Types.NielsenCBET

Description

Documentation

data NielsenCBET

Nielsen CBET

See: newNielsenCBET smart constructor.

Constructors

NielsenCBET'

Fields

cbetCheckDigitString :: Text

Enter the CBET check digits to use in the watermark.

cbetStepaside :: NielsenWatermarksCbetStepaside

Determines the method of CBET insertion mode when prior encoding is detected on the same layer.

csid :: Text

Enter the CBET Source ID (CSID) to use in the watermark

newNielsenCBET

Arguments

:: Text
$sel:cbetCheckDigitString:NielsenCBET'

-> NielsenWatermarksCbetStepaside
$sel:cbetStepaside:NielsenCBET'

-> Text
$sel:csid:NielsenCBET'

-> NielsenCBET

Create a value of NielsenCBET with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:cbetCheckDigitString:NielsenCBET', nielsenCBET_cbetCheckDigitString - Enter the CBET check digits to use in the watermark.

$sel:cbetStepaside:NielsenCBET', nielsenCBET_cbetStepaside - Determines the method of CBET insertion mode when prior encoding is detected on the same layer.

$sel:csid:NielsenCBET', nielsenCBET_csid - Enter the CBET Source ID (CSID) to use in the watermark

nielsenCBET_cbetCheckDigitString :: Lens' NielsenCBET Text

Enter the CBET check digits to use in the watermark.

nielsenCBET_cbetStepaside :: Lens' NielsenCBET NielsenWatermarksCbetStepaside

Determines the method of CBET insertion mode when prior encoding is detected on the same layer.

nielsenCBET_csid :: Lens' NielsenCBET Text

Enter the CBET Source ID (CSID) to use in the watermark