Amazonka.MediaLive.Types.WavSettings

Description

Documentation

data WavSettings

Wav Settings

See: newWavSettings smart constructor.

Constructors

WavSettings'

Fields

bitDepth :: Maybe Double

Bits per sample.

codingMode :: Maybe WavCodingMode

The audio coding mode for the WAV audio. The mode determines the number of channels in the audio.

sampleRate :: Maybe Double

Sample rate in Hz.

newWavSettings :: WavSettings

Create a value of WavSettings with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:bitDepth:WavSettings', wavSettings_bitDepth - Bits per sample.

$sel:codingMode:WavSettings', wavSettings_codingMode - The audio coding mode for the WAV audio. The mode determines the number of channels in the audio.

$sel:sampleRate:WavSettings', wavSettings_sampleRate - Sample rate in Hz.

wavSettings_bitDepth :: Lens' WavSettings (Maybe Double)

Bits per sample.

wavSettings_codingMode :: Lens' WavSettings (Maybe WavCodingMode)

The audio coding mode for the WAV audio. The mode determines the number of channels in the audio.

wavSettings_sampleRate :: Lens' WavSettings (Maybe Double)

Sample rate in Hz.