Amazonka.MediaPackage.CreateOriginEndpoint

Description

Creates a new OriginEndpoint record.

Creating a Request

data CreateOriginEndpoint

Configuration parameters used to create a new OriginEndpoint.

See: newCreateOriginEndpoint smart constructor.

Constructors

CreateOriginEndpoint'

Fields

whitelist :: Maybe [Text]

A list of source IP CIDR blocks that will be allowed to access the OriginEndpoint.

hlsPackage :: Maybe HlsPackage

manifestName :: Maybe Text

A short string that will be used as the filename of the OriginEndpoint URL (defaults to "index").

authorization :: Maybe Authorization

startoverWindowSeconds :: Maybe Int

Maximum duration (seconds) of content to retain for startover playback. If not specified, startover playback will be disabled for the OriginEndpoint.

dashPackage :: Maybe DashPackage

mssPackage :: Maybe MssPackage

timeDelaySeconds :: Maybe Int

Amount of delay (seconds) to enforce on the playback of live content. If not specified, there will be no time delay in effect for the OriginEndpoint.

cmafPackage :: Maybe CmafPackageCreateOrUpdateParameters

description :: Maybe Text

A short text description of the OriginEndpoint.

tags :: Maybe (HashMap Text Text)

origination :: Maybe Origination

Control whether origination of video is allowed for this OriginEndpoint. If set to ALLOW, the OriginEndpoint may by requested, pursuant to any other form of access control. If set to DENY, the OriginEndpoint may not be requested. This can be helpful for Live to VOD harvesting, or for temporarily disabling origination

channelId :: Text

The ID of the Channel that the OriginEndpoint will be associated with. This cannot be changed after the OriginEndpoint is created.

id :: Text

The ID of the OriginEndpoint. The ID must be unique within the region and it cannot be changed after the OriginEndpoint is created.

newCreateOriginEndpoint

Arguments

:: Text
$sel:channelId:CreateOriginEndpoint'

-> Text
$sel:id:CreateOriginEndpoint'

-> CreateOriginEndpoint

Create a value of CreateOriginEndpoint with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:whitelist:CreateOriginEndpoint', createOriginEndpoint_whitelist - A list of source IP CIDR blocks that will be allowed to access the OriginEndpoint.

$sel:hlsPackage:CreateOriginEndpoint', createOriginEndpoint_hlsPackage - Undocumented member.

$sel:manifestName:CreateOriginEndpoint', createOriginEndpoint_manifestName - A short string that will be used as the filename of the OriginEndpoint URL (defaults to "index").

$sel:authorization:CreateOriginEndpoint', createOriginEndpoint_authorization - Undocumented member.

$sel:startoverWindowSeconds:CreateOriginEndpoint', createOriginEndpoint_startoverWindowSeconds - Maximum duration (seconds) of content to retain for startover playback. If not specified, startover playback will be disabled for the OriginEndpoint.

$sel:dashPackage:CreateOriginEndpoint', createOriginEndpoint_dashPackage - Undocumented member.

$sel:mssPackage:CreateOriginEndpoint', createOriginEndpoint_mssPackage - Undocumented member.

$sel:timeDelaySeconds:CreateOriginEndpoint', createOriginEndpoint_timeDelaySeconds - Amount of delay (seconds) to enforce on the playback of live content. If not specified, there will be no time delay in effect for the OriginEndpoint.

$sel:cmafPackage:CreateOriginEndpoint', createOriginEndpoint_cmafPackage - Undocumented member.

$sel:description:CreateOriginEndpoint', createOriginEndpoint_description - A short text description of the OriginEndpoint.

$sel:tags:CreateOriginEndpoint', createOriginEndpoint_tags - Undocumented member.

$sel:origination:CreateOriginEndpoint', createOriginEndpoint_origination - Control whether origination of video is allowed for this OriginEndpoint. If set to ALLOW, the OriginEndpoint may by requested, pursuant to any other form of access control. If set to DENY, the OriginEndpoint may not be requested. This can be helpful for Live to VOD harvesting, or for temporarily disabling origination

$sel:channelId:CreateOriginEndpoint', createOriginEndpoint_channelId - The ID of the Channel that the OriginEndpoint will be associated with. This cannot be changed after the OriginEndpoint is created.

$sel:id:CreateOriginEndpoint', createOriginEndpoint_id - The ID of the OriginEndpoint. The ID must be unique within the region and it cannot be changed after the OriginEndpoint is created.

Request Lenses

createOriginEndpoint_whitelist :: Lens' CreateOriginEndpoint (Maybe [Text])

A list of source IP CIDR blocks that will be allowed to access the OriginEndpoint.

createOriginEndpoint_hlsPackage :: Lens' CreateOriginEndpoint (Maybe HlsPackage)

Undocumented member.

createOriginEndpoint_manifestName :: Lens' CreateOriginEndpoint (Maybe Text)

A short string that will be used as the filename of the OriginEndpoint URL (defaults to "index").

createOriginEndpoint_authorization :: Lens' CreateOriginEndpoint (Maybe Authorization)

Undocumented member.

createOriginEndpoint_startoverWindowSeconds :: Lens' CreateOriginEndpoint (Maybe Int)

Maximum duration (seconds) of content to retain for startover playback. If not specified, startover playback will be disabled for the OriginEndpoint.

createOriginEndpoint_dashPackage :: Lens' CreateOriginEndpoint (Maybe DashPackage)

Undocumented member.

createOriginEndpoint_mssPackage :: Lens' CreateOriginEndpoint (Maybe MssPackage)

Undocumented member.

createOriginEndpoint_timeDelaySeconds :: Lens' CreateOriginEndpoint (Maybe Int)

Amount of delay (seconds) to enforce on the playback of live content. If not specified, there will be no time delay in effect for the OriginEndpoint.

createOriginEndpoint_cmafPackage :: Lens' CreateOriginEndpoint (Maybe CmafPackageCreateOrUpdateParameters)

Undocumented member.

createOriginEndpoint_description :: Lens' CreateOriginEndpoint (Maybe Text)

A short text description of the OriginEndpoint.

createOriginEndpoint_tags :: Lens' CreateOriginEndpoint (Maybe (HashMap Text Text))

Undocumented member.

createOriginEndpoint_origination :: Lens' CreateOriginEndpoint (Maybe Origination)

Control whether origination of video is allowed for this OriginEndpoint. If set to ALLOW, the OriginEndpoint may by requested, pursuant to any other form of access control. If set to DENY, the OriginEndpoint may not be requested. This can be helpful for Live to VOD harvesting, or for temporarily disabling origination

createOriginEndpoint_channelId :: Lens' CreateOriginEndpoint Text

The ID of the Channel that the OriginEndpoint will be associated with. This cannot be changed after the OriginEndpoint is created.

createOriginEndpoint_id :: Lens' CreateOriginEndpoint Text

The ID of the OriginEndpoint. The ID must be unique within the region and it cannot be changed after the OriginEndpoint is created.

Destructuring the Response

data CreateOriginEndpointResponse

See: newCreateOriginEndpointResponse smart constructor.

Constructors

CreateOriginEndpointResponse'

Fields

whitelist :: Maybe [Text]

A list of source IP CIDR blocks that will be allowed to access the OriginEndpoint.

hlsPackage :: Maybe HlsPackage

arn :: Maybe Text

The Amazon Resource Name (ARN) assigned to the OriginEndpoint.

manifestName :: Maybe Text

A short string appended to the end of the OriginEndpoint URL.

url :: Maybe Text

The URL of the packaged OriginEndpoint for consumption.

authorization :: Maybe Authorization

channelId :: Maybe Text

The ID of the Channel the OriginEndpoint is associated with.

startoverWindowSeconds :: Maybe Int

Maximum duration (seconds) of content to retain for startover playback. If not specified, startover playback will be disabled for the OriginEndpoint.

dashPackage :: Maybe DashPackage

mssPackage :: Maybe MssPackage

id :: Maybe Text

The ID of the OriginEndpoint.

timeDelaySeconds :: Maybe Int

Amount of delay (seconds) to enforce on the playback of live content. If not specified, there will be no time delay in effect for the OriginEndpoint.

cmafPackage :: Maybe CmafPackage

description :: Maybe Text

A short text description of the OriginEndpoint.

tags :: Maybe (HashMap Text Text)

origination :: Maybe Origination

Control whether origination of video is allowed for this OriginEndpoint. If set to ALLOW, the OriginEndpoint may by requested, pursuant to any other form of access control. If set to DENY, the OriginEndpoint may not be requested. This can be helpful for Live to VOD harvesting, or for temporarily disabling origination

httpStatus :: Int

The response's http status code.

newCreateOriginEndpointResponse

Arguments

:: Int
$sel:httpStatus:CreateOriginEndpointResponse'

-> CreateOriginEndpointResponse

Create a value of CreateOriginEndpointResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:whitelist:CreateOriginEndpoint', createOriginEndpointResponse_whitelist - A list of source IP CIDR blocks that will be allowed to access the OriginEndpoint.

$sel:hlsPackage:CreateOriginEndpoint', createOriginEndpointResponse_hlsPackage - Undocumented member.

$sel:arn:CreateOriginEndpointResponse', createOriginEndpointResponse_arn - The Amazon Resource Name (ARN) assigned to the OriginEndpoint.

$sel:manifestName:CreateOriginEndpoint', createOriginEndpointResponse_manifestName - A short string appended to the end of the OriginEndpoint URL.

$sel:url:CreateOriginEndpointResponse', createOriginEndpointResponse_url - The URL of the packaged OriginEndpoint for consumption.

$sel:authorization:CreateOriginEndpoint', createOriginEndpointResponse_authorization - Undocumented member.

$sel:channelId:CreateOriginEndpoint', createOriginEndpointResponse_channelId - The ID of the Channel the OriginEndpoint is associated with.

$sel:startoverWindowSeconds:CreateOriginEndpoint', createOriginEndpointResponse_startoverWindowSeconds - Maximum duration (seconds) of content to retain for startover playback. If not specified, startover playback will be disabled for the OriginEndpoint.

$sel:dashPackage:CreateOriginEndpoint', createOriginEndpointResponse_dashPackage - Undocumented member.

$sel:mssPackage:CreateOriginEndpoint', createOriginEndpointResponse_mssPackage - Undocumented member.

$sel:id:CreateOriginEndpoint', createOriginEndpointResponse_id - The ID of the OriginEndpoint.

$sel:timeDelaySeconds:CreateOriginEndpoint', createOriginEndpointResponse_timeDelaySeconds - Amount of delay (seconds) to enforce on the playback of live content. If not specified, there will be no time delay in effect for the OriginEndpoint.

$sel:cmafPackage:CreateOriginEndpoint', createOriginEndpointResponse_cmafPackage - Undocumented member.

$sel:description:CreateOriginEndpoint', createOriginEndpointResponse_description - A short text description of the OriginEndpoint.

$sel:tags:CreateOriginEndpoint', createOriginEndpointResponse_tags - Undocumented member.

$sel:origination:CreateOriginEndpoint', createOriginEndpointResponse_origination - Control whether origination of video is allowed for this OriginEndpoint. If set to ALLOW, the OriginEndpoint may by requested, pursuant to any other form of access control. If set to DENY, the OriginEndpoint may not be requested. This can be helpful for Live to VOD harvesting, or for temporarily disabling origination

$sel:httpStatus:CreateOriginEndpointResponse', createOriginEndpointResponse_httpStatus - The response's http status code.

Response Lenses

createOriginEndpointResponse_whitelist :: Lens' CreateOriginEndpointResponse (Maybe [Text])

A list of source IP CIDR blocks that will be allowed to access the OriginEndpoint.

createOriginEndpointResponse_hlsPackage :: Lens' CreateOriginEndpointResponse (Maybe HlsPackage)

Undocumented member.

createOriginEndpointResponse_arn :: Lens' CreateOriginEndpointResponse (Maybe Text)

The Amazon Resource Name (ARN) assigned to the OriginEndpoint.

createOriginEndpointResponse_manifestName :: Lens' CreateOriginEndpointResponse (Maybe Text)

A short string appended to the end of the OriginEndpoint URL.

createOriginEndpointResponse_url :: Lens' CreateOriginEndpointResponse (Maybe Text)

The URL of the packaged OriginEndpoint for consumption.

createOriginEndpointResponse_authorization :: Lens' CreateOriginEndpointResponse (Maybe Authorization)

Undocumented member.

createOriginEndpointResponse_channelId :: Lens' CreateOriginEndpointResponse (Maybe Text)

The ID of the Channel the OriginEndpoint is associated with.

createOriginEndpointResponse_startoverWindowSeconds :: Lens' CreateOriginEndpointResponse (Maybe Int)

Maximum duration (seconds) of content to retain for startover playback. If not specified, startover playback will be disabled for the OriginEndpoint.

createOriginEndpointResponse_dashPackage :: Lens' CreateOriginEndpointResponse (Maybe DashPackage)

Undocumented member.

createOriginEndpointResponse_mssPackage :: Lens' CreateOriginEndpointResponse (Maybe MssPackage)

Undocumented member.

createOriginEndpointResponse_id :: Lens' CreateOriginEndpointResponse (Maybe Text)

The ID of the OriginEndpoint.

createOriginEndpointResponse_timeDelaySeconds :: Lens' CreateOriginEndpointResponse (Maybe Int)

Amount of delay (seconds) to enforce on the playback of live content. If not specified, there will be no time delay in effect for the OriginEndpoint.

createOriginEndpointResponse_cmafPackage :: Lens' CreateOriginEndpointResponse (Maybe CmafPackage)

Undocumented member.

createOriginEndpointResponse_description :: Lens' CreateOriginEndpointResponse (Maybe Text)

A short text description of the OriginEndpoint.

createOriginEndpointResponse_tags :: Lens' CreateOriginEndpointResponse (Maybe (HashMap Text Text))

Undocumented member.

createOriginEndpointResponse_origination :: Lens' CreateOriginEndpointResponse (Maybe Origination)

Control whether origination of video is allowed for this OriginEndpoint. If set to ALLOW, the OriginEndpoint may by requested, pursuant to any other form of access control. If set to DENY, the OriginEndpoint may not be requested. This can be helpful for Live to VOD harvesting, or for temporarily disabling origination

createOriginEndpointResponse_httpStatus :: Lens' CreateOriginEndpointResponse Int

The response's http status code.