Amazonka.MediaPackage.DescribeHarvestJob

Description

Gets details about an existing HarvestJob.

Creating a Request

data DescribeHarvestJob

See: newDescribeHarvestJob smart constructor.

Constructors

DescribeHarvestJob'

Fields

id :: Text

The ID of the HarvestJob.

newDescribeHarvestJob

Arguments

:: Text
$sel:id:DescribeHarvestJob'

-> DescribeHarvestJob

Create a value of DescribeHarvestJob with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:id:DescribeHarvestJob', describeHarvestJob_id - The ID of the HarvestJob.

Request Lenses

describeHarvestJob_id :: Lens' DescribeHarvestJob Text

The ID of the HarvestJob.

Destructuring the Response

data DescribeHarvestJobResponse

See: newDescribeHarvestJobResponse smart constructor.

Constructors

DescribeHarvestJobResponse'

Fields

status :: Maybe Status

The current status of the HarvestJob. Consider setting up a CloudWatch Event to listen for HarvestJobs as they succeed or fail. In the event of failure, the CloudWatch Event will include an explanation of why the HarvestJob failed.

originEndpointId :: Maybe Text

The ID of the OriginEndpoint that the HarvestJob will harvest from. This cannot be changed after the HarvestJob is submitted.

startTime :: Maybe Text

The start of the time-window which will be harvested.

arn :: Maybe Text

The Amazon Resource Name (ARN) assigned to the HarvestJob.

createdAt :: Maybe Text

The time the HarvestJob was submitted

channelId :: Maybe Text

The ID of the Channel that the HarvestJob will harvest from.

s3Destination :: Maybe S3Destination

endTime :: Maybe Text

The end of the time-window which will be harvested.

id :: Maybe Text

The ID of the HarvestJob. The ID must be unique within the region and it cannot be changed after the HarvestJob is submitted.

httpStatus :: Int

The response's http status code.

newDescribeHarvestJobResponse

Arguments

:: Int
$sel:httpStatus:DescribeHarvestJobResponse'

-> DescribeHarvestJobResponse

Create a value of DescribeHarvestJobResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:status:DescribeHarvestJobResponse', describeHarvestJobResponse_status - The current status of the HarvestJob. Consider setting up a CloudWatch Event to listen for HarvestJobs as they succeed or fail. In the event of failure, the CloudWatch Event will include an explanation of why the HarvestJob failed.

$sel:originEndpointId:DescribeHarvestJobResponse', describeHarvestJobResponse_originEndpointId - The ID of the OriginEndpoint that the HarvestJob will harvest from. This cannot be changed after the HarvestJob is submitted.

$sel:startTime:DescribeHarvestJobResponse', describeHarvestJobResponse_startTime - The start of the time-window which will be harvested.

$sel:arn:DescribeHarvestJobResponse', describeHarvestJobResponse_arn - The Amazon Resource Name (ARN) assigned to the HarvestJob.

$sel:createdAt:DescribeHarvestJobResponse', describeHarvestJobResponse_createdAt - The time the HarvestJob was submitted

$sel:channelId:DescribeHarvestJobResponse', describeHarvestJobResponse_channelId - The ID of the Channel that the HarvestJob will harvest from.

$sel:s3Destination:DescribeHarvestJobResponse', describeHarvestJobResponse_s3Destination - Undocumented member.

$sel:endTime:DescribeHarvestJobResponse', describeHarvestJobResponse_endTime - The end of the time-window which will be harvested.

$sel:id:DescribeHarvestJob', describeHarvestJobResponse_id - The ID of the HarvestJob. The ID must be unique within the region and it cannot be changed after the HarvestJob is submitted.

$sel:httpStatus:DescribeHarvestJobResponse', describeHarvestJobResponse_httpStatus - The response's http status code.

Response Lenses

describeHarvestJobResponse_status :: Lens' DescribeHarvestJobResponse (Maybe Status)

The current status of the HarvestJob. Consider setting up a CloudWatch Event to listen for HarvestJobs as they succeed or fail. In the event of failure, the CloudWatch Event will include an explanation of why the HarvestJob failed.

describeHarvestJobResponse_originEndpointId :: Lens' DescribeHarvestJobResponse (Maybe Text)

The ID of the OriginEndpoint that the HarvestJob will harvest from. This cannot be changed after the HarvestJob is submitted.

describeHarvestJobResponse_startTime :: Lens' DescribeHarvestJobResponse (Maybe Text)

The start of the time-window which will be harvested.

describeHarvestJobResponse_arn :: Lens' DescribeHarvestJobResponse (Maybe Text)

The Amazon Resource Name (ARN) assigned to the HarvestJob.

describeHarvestJobResponse_createdAt :: Lens' DescribeHarvestJobResponse (Maybe Text)

The time the HarvestJob was submitted

describeHarvestJobResponse_channelId :: Lens' DescribeHarvestJobResponse (Maybe Text)

The ID of the Channel that the HarvestJob will harvest from.

describeHarvestJobResponse_s3Destination :: Lens' DescribeHarvestJobResponse (Maybe S3Destination)

Undocumented member.

describeHarvestJobResponse_endTime :: Lens' DescribeHarvestJobResponse (Maybe Text)

The end of the time-window which will be harvested.

describeHarvestJobResponse_id :: Lens' DescribeHarvestJobResponse (Maybe Text)

The ID of the HarvestJob. The ID must be unique within the region and it cannot be changed after the HarvestJob is submitted.

describeHarvestJobResponse_httpStatus :: Lens' DescribeHarvestJobResponse Int

The response's http status code.