Amazonka.MediaPackage.Types.DashPackage

Description

Documentation

data DashPackage

A Dynamic Adaptive Streaming over HTTP (DASH) packaging configuration.

See: newDashPackage smart constructor.

Constructors

DashPackage'

Fields

adsOnDeliveryRestrictions :: Maybe AdsOnDeliveryRestrictions

minBufferTimeSeconds :: Maybe Int

Minimum duration (in seconds) that a player will buffer media before starting the presentation.

utcTiming :: Maybe UtcTiming

Determines the type of UTCTiming included in the Media Presentation Description (MPD)

segmentTemplateFormat :: Maybe SegmentTemplateFormat

Determines the type of SegmentTemplate included in the Media Presentation Description (MPD). When set to NUMBER_WITH_TIMELINE, a full timeline is presented in each SegmentTemplate, with $Number$ media URLs. When set to TIME_WITH_TIMELINE, a full timeline is presented in each SegmentTemplate, with $Time$ media URLs. When set to NUMBER_WITH_DURATION, only a duration is included in each SegmentTemplate, with $Number$ media URLs.

profile :: Maybe Profile

The Dynamic Adaptive Streaming over HTTP (DASH) profile type. When set to "HBBTV_1_5", HbbTV 1.5 compliant output is enabled.

segmentDurationSeconds :: Maybe Int

Duration (in seconds) of each segment. Actual segments will be rounded to the nearest multiple of the source segment duration.

utcTimingUri :: Maybe Text

Specifies the value attribute of the UTCTiming field when utcTiming is set to HTTP-ISO or HTTP-HEAD

streamSelection :: Maybe StreamSelection

encryption :: Maybe DashEncryption

minUpdatePeriodSeconds :: Maybe Int

Minimum duration (in seconds) between potential changes to the Dynamic Adaptive Streaming over HTTP (DASH) Media Presentation Description (MPD).

manifestLayout :: Maybe ManifestLayout

Determines the position of some tags in the Media Presentation Description (MPD). When set to FULL, elements like SegmentTemplate and ContentProtection are included in each Representation. When set to COMPACT, duplicate elements are combined and presented at the AdaptationSet level.

suggestedPresentationDelaySeconds :: Maybe Int

Duration (in seconds) to delay live content before presentation.

manifestWindowSeconds :: Maybe Int

Time window (in seconds) contained in each manifest.

adTriggers :: Maybe [AdTriggersElement]

periodTriggers :: Maybe [PeriodTriggersElement]

A list of triggers that controls when the outgoing Dynamic Adaptive Streaming over HTTP (DASH) Media Presentation Description (MPD) will be partitioned into multiple periods. If empty, the content will not be partitioned into more than one period. If the list contains "ADS", new periods will be created where the Channel source contains SCTE-35 ad markers.

newDashPackage :: DashPackage

Create a value of DashPackage with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:adsOnDeliveryRestrictions:DashPackage', dashPackage_adsOnDeliveryRestrictions - Undocumented member.

$sel:minBufferTimeSeconds:DashPackage', dashPackage_minBufferTimeSeconds - Minimum duration (in seconds) that a player will buffer media before starting the presentation.

$sel:utcTiming:DashPackage', dashPackage_utcTiming - Determines the type of UTCTiming included in the Media Presentation Description (MPD)

$sel:segmentTemplateFormat:DashPackage', dashPackage_segmentTemplateFormat - Determines the type of SegmentTemplate included in the Media Presentation Description (MPD). When set to NUMBER_WITH_TIMELINE, a full timeline is presented in each SegmentTemplate, with $Number$ media URLs. When set to TIME_WITH_TIMELINE, a full timeline is presented in each SegmentTemplate, with $Time$ media URLs. When set to NUMBER_WITH_DURATION, only a duration is included in each SegmentTemplate, with $Number$ media URLs.

$sel:profile:DashPackage', dashPackage_profile - The Dynamic Adaptive Streaming over HTTP (DASH) profile type. When set to "HBBTV_1_5", HbbTV 1.5 compliant output is enabled.

$sel:segmentDurationSeconds:DashPackage', dashPackage_segmentDurationSeconds - Duration (in seconds) of each segment. Actual segments will be rounded to the nearest multiple of the source segment duration.

$sel:utcTimingUri:DashPackage', dashPackage_utcTimingUri - Specifies the value attribute of the UTCTiming field when utcTiming is set to HTTP-ISO or HTTP-HEAD

$sel:streamSelection:DashPackage', dashPackage_streamSelection - Undocumented member.

$sel:encryption:DashPackage', dashPackage_encryption - Undocumented member.

$sel:minUpdatePeriodSeconds:DashPackage', dashPackage_minUpdatePeriodSeconds - Minimum duration (in seconds) between potential changes to the Dynamic Adaptive Streaming over HTTP (DASH) Media Presentation Description (MPD).

$sel:manifestLayout:DashPackage', dashPackage_manifestLayout - Determines the position of some tags in the Media Presentation Description (MPD). When set to FULL, elements like SegmentTemplate and ContentProtection are included in each Representation. When set to COMPACT, duplicate elements are combined and presented at the AdaptationSet level.

$sel:suggestedPresentationDelaySeconds:DashPackage', dashPackage_suggestedPresentationDelaySeconds - Duration (in seconds) to delay live content before presentation.

$sel:manifestWindowSeconds:DashPackage', dashPackage_manifestWindowSeconds - Time window (in seconds) contained in each manifest.

$sel:adTriggers:DashPackage', dashPackage_adTriggers - Undocumented member.

$sel:periodTriggers:DashPackage', dashPackage_periodTriggers - A list of triggers that controls when the outgoing Dynamic Adaptive Streaming over HTTP (DASH) Media Presentation Description (MPD) will be partitioned into multiple periods. If empty, the content will not be partitioned into more than one period. If the list contains "ADS", new periods will be created where the Channel source contains SCTE-35 ad markers.

dashPackage_adsOnDeliveryRestrictions :: Lens' DashPackage (Maybe AdsOnDeliveryRestrictions)

Undocumented member.

dashPackage_minBufferTimeSeconds :: Lens' DashPackage (Maybe Int)

Minimum duration (in seconds) that a player will buffer media before starting the presentation.

dashPackage_utcTiming :: Lens' DashPackage (Maybe UtcTiming)

Determines the type of UTCTiming included in the Media Presentation Description (MPD)

dashPackage_segmentTemplateFormat :: Lens' DashPackage (Maybe SegmentTemplateFormat)

Determines the type of SegmentTemplate included in the Media Presentation Description (MPD). When set to NUMBER_WITH_TIMELINE, a full timeline is presented in each SegmentTemplate, with $Number$ media URLs. When set to TIME_WITH_TIMELINE, a full timeline is presented in each SegmentTemplate, with $Time$ media URLs. When set to NUMBER_WITH_DURATION, only a duration is included in each SegmentTemplate, with $Number$ media URLs.

dashPackage_profile :: Lens' DashPackage (Maybe Profile)

The Dynamic Adaptive Streaming over HTTP (DASH) profile type. When set to "HBBTV_1_5", HbbTV 1.5 compliant output is enabled.

dashPackage_segmentDurationSeconds :: Lens' DashPackage (Maybe Int)

Duration (in seconds) of each segment. Actual segments will be rounded to the nearest multiple of the source segment duration.

dashPackage_utcTimingUri :: Lens' DashPackage (Maybe Text)

Specifies the value attribute of the UTCTiming field when utcTiming is set to HTTP-ISO or HTTP-HEAD

dashPackage_streamSelection :: Lens' DashPackage (Maybe StreamSelection)

Undocumented member.

dashPackage_encryption :: Lens' DashPackage (Maybe DashEncryption)

Undocumented member.

dashPackage_minUpdatePeriodSeconds :: Lens' DashPackage (Maybe Int)

Minimum duration (in seconds) between potential changes to the Dynamic Adaptive Streaming over HTTP (DASH) Media Presentation Description (MPD).

dashPackage_manifestLayout :: Lens' DashPackage (Maybe ManifestLayout)

Determines the position of some tags in the Media Presentation Description (MPD). When set to FULL, elements like SegmentTemplate and ContentProtection are included in each Representation. When set to COMPACT, duplicate elements are combined and presented at the AdaptationSet level.

dashPackage_suggestedPresentationDelaySeconds :: Lens' DashPackage (Maybe Int)

Duration (in seconds) to delay live content before presentation.

dashPackage_manifestWindowSeconds :: Lens' DashPackage (Maybe Int)

Time window (in seconds) contained in each manifest.

dashPackage_adTriggers :: Lens' DashPackage (Maybe [AdTriggersElement])

Undocumented member.

dashPackage_periodTriggers :: Lens' DashPackage (Maybe [PeriodTriggersElement])

A list of triggers that controls when the outgoing Dynamic Adaptive Streaming over HTTP (DASH) Media Presentation Description (MPD) will be partitioned into multiple periods. If empty, the content will not be partitioned into more than one period. If the list contains "ADS", new periods will be created where the Channel source contains SCTE-35 ad markers.