Amazonka.MediaPackage.Types.EncryptionContractConfiguration

Description

Documentation

data EncryptionContractConfiguration

Use encryptionContractConfiguration to configure one or more content encryption keys for your endpoints that use SPEKE 2.0. The encryption contract defines which content keys are used to encrypt the audio and video tracks in your stream. To configure the encryption contract, specify which audio and video encryption presets to use. Note the following considerations when using encryptionContractConfiguration: encryptionContractConfiguration can be used for DASH or CMAF endpoints that use SPEKE 2.0. SPEKE 2.0 relies on the CPIX 2.3 specification. You must disable key rotation for this endpoint by setting keyRotationIntervalSeconds to 0.

See: newEncryptionContractConfiguration smart constructor.

Constructors

EncryptionContractConfiguration'

Fields

presetSpeke20Audio :: PresetSpeke20Audio

A collection of audio encryption presets.

presetSpeke20Video :: PresetSpeke20Video

A collection of video encryption presets.

newEncryptionContractConfiguration

Arguments

:: PresetSpeke20Audio
$sel:presetSpeke20Audio:EncryptionContractConfiguration'

-> PresetSpeke20Video
$sel:presetSpeke20Video:EncryptionContractConfiguration'

-> EncryptionContractConfiguration

Create a value of EncryptionContractConfiguration with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:presetSpeke20Audio:EncryptionContractConfiguration', encryptionContractConfiguration_presetSpeke20Audio - A collection of audio encryption presets.

$sel:presetSpeke20Video:EncryptionContractConfiguration', encryptionContractConfiguration_presetSpeke20Video - A collection of video encryption presets.

encryptionContractConfiguration_presetSpeke20Audio :: Lens' EncryptionContractConfiguration PresetSpeke20Audio

A collection of audio encryption presets.

encryptionContractConfiguration_presetSpeke20Video :: Lens' EncryptionContractConfiguration PresetSpeke20Video

A collection of video encryption presets.