Amazonka.MediaPackage.Types.OriginEndpoint

Description

Documentation

data OriginEndpoint

An OriginEndpoint resource configuration.

See: newOriginEndpoint smart constructor.

Constructors

OriginEndpoint'

Fields

whitelist :: Maybe [Text]

A list of source IP CIDR blocks that will be allowed to access the OriginEndpoint.

hlsPackage :: Maybe HlsPackage

arn :: Maybe Text

The Amazon Resource Name (ARN) assigned to the OriginEndpoint.

manifestName :: Maybe Text

A short string appended to the end of the OriginEndpoint URL.

url :: Maybe Text

The URL of the packaged OriginEndpoint for consumption.

authorization :: Maybe Authorization

channelId :: Maybe Text

The ID of the Channel the OriginEndpoint is associated with.

startoverWindowSeconds :: Maybe Int

Maximum duration (seconds) of content to retain for startover playback. If not specified, startover playback will be disabled for the OriginEndpoint.

dashPackage :: Maybe DashPackage

mssPackage :: Maybe MssPackage

id :: Maybe Text

The ID of the OriginEndpoint.

timeDelaySeconds :: Maybe Int

Amount of delay (seconds) to enforce on the playback of live content. If not specified, there will be no time delay in effect for the OriginEndpoint.

cmafPackage :: Maybe CmafPackage

description :: Maybe Text

A short text description of the OriginEndpoint.

tags :: Maybe (HashMap Text Text)

origination :: Maybe Origination

Control whether origination of video is allowed for this OriginEndpoint. If set to ALLOW, the OriginEndpoint may by requested, pursuant to any other form of access control. If set to DENY, the OriginEndpoint may not be requested. This can be helpful for Live to VOD harvesting, or for temporarily disabling origination

newOriginEndpoint :: OriginEndpoint

Create a value of OriginEndpoint with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:whitelist:OriginEndpoint', originEndpoint_whitelist - A list of source IP CIDR blocks that will be allowed to access the OriginEndpoint.

$sel:hlsPackage:OriginEndpoint', originEndpoint_hlsPackage - Undocumented member.

$sel:arn:OriginEndpoint', originEndpoint_arn - The Amazon Resource Name (ARN) assigned to the OriginEndpoint.

$sel:manifestName:OriginEndpoint', originEndpoint_manifestName - A short string appended to the end of the OriginEndpoint URL.

$sel:url:OriginEndpoint', originEndpoint_url - The URL of the packaged OriginEndpoint for consumption.

$sel:authorization:OriginEndpoint', originEndpoint_authorization - Undocumented member.

$sel:channelId:OriginEndpoint', originEndpoint_channelId - The ID of the Channel the OriginEndpoint is associated with.

$sel:startoverWindowSeconds:OriginEndpoint', originEndpoint_startoverWindowSeconds - Maximum duration (seconds) of content to retain for startover playback. If not specified, startover playback will be disabled for the OriginEndpoint.

$sel:dashPackage:OriginEndpoint', originEndpoint_dashPackage - Undocumented member.

$sel:mssPackage:OriginEndpoint', originEndpoint_mssPackage - Undocumented member.

$sel:id:OriginEndpoint', originEndpoint_id - The ID of the OriginEndpoint.

$sel:timeDelaySeconds:OriginEndpoint', originEndpoint_timeDelaySeconds - Amount of delay (seconds) to enforce on the playback of live content. If not specified, there will be no time delay in effect for the OriginEndpoint.

$sel:cmafPackage:OriginEndpoint', originEndpoint_cmafPackage - Undocumented member.

$sel:description:OriginEndpoint', originEndpoint_description - A short text description of the OriginEndpoint.

$sel:tags:OriginEndpoint', originEndpoint_tags - Undocumented member.

$sel:origination:OriginEndpoint', originEndpoint_origination - Control whether origination of video is allowed for this OriginEndpoint. If set to ALLOW, the OriginEndpoint may by requested, pursuant to any other form of access control. If set to DENY, the OriginEndpoint may not be requested. This can be helpful for Live to VOD harvesting, or for temporarily disabling origination

originEndpoint_whitelist :: Lens' OriginEndpoint (Maybe [Text])

A list of source IP CIDR blocks that will be allowed to access the OriginEndpoint.

originEndpoint_hlsPackage :: Lens' OriginEndpoint (Maybe HlsPackage)

Undocumented member.

originEndpoint_arn :: Lens' OriginEndpoint (Maybe Text)

The Amazon Resource Name (ARN) assigned to the OriginEndpoint.

originEndpoint_manifestName :: Lens' OriginEndpoint (Maybe Text)

A short string appended to the end of the OriginEndpoint URL.

originEndpoint_url :: Lens' OriginEndpoint (Maybe Text)

The URL of the packaged OriginEndpoint for consumption.

originEndpoint_authorization :: Lens' OriginEndpoint (Maybe Authorization)

Undocumented member.

originEndpoint_channelId :: Lens' OriginEndpoint (Maybe Text)

The ID of the Channel the OriginEndpoint is associated with.

originEndpoint_startoverWindowSeconds :: Lens' OriginEndpoint (Maybe Int)

Maximum duration (seconds) of content to retain for startover playback. If not specified, startover playback will be disabled for the OriginEndpoint.

originEndpoint_dashPackage :: Lens' OriginEndpoint (Maybe DashPackage)

Undocumented member.

originEndpoint_mssPackage :: Lens' OriginEndpoint (Maybe MssPackage)

Undocumented member.

originEndpoint_id :: Lens' OriginEndpoint (Maybe Text)

The ID of the OriginEndpoint.

originEndpoint_timeDelaySeconds :: Lens' OriginEndpoint (Maybe Int)

Amount of delay (seconds) to enforce on the playback of live content. If not specified, there will be no time delay in effect for the OriginEndpoint.

originEndpoint_cmafPackage :: Lens' OriginEndpoint (Maybe CmafPackage)

Undocumented member.

originEndpoint_description :: Lens' OriginEndpoint (Maybe Text)

A short text description of the OriginEndpoint.

originEndpoint_tags :: Lens' OriginEndpoint (Maybe (HashMap Text Text))

Undocumented member.

originEndpoint_origination :: Lens' OriginEndpoint (Maybe Origination)

Control whether origination of video is allowed for this OriginEndpoint. If set to ALLOW, the OriginEndpoint may by requested, pursuant to any other form of access control. If set to DENY, the OriginEndpoint may not be requested. This can be helpful for Live to VOD harvesting, or for temporarily disabling origination