Amazonka.MediaPackageVOD

Description

Derived from API version 2018-11-07 of the AWS service descriptions, licensed under Apache 2.0.

AWS Elemental MediaPackage VOD

Service Configuration

defaultService :: Service

API version 2018-11-07 of the Amazon Elemental MediaPackage VOD SDK configuration.

Errors

Error matchers are designed for use with the functions provided by Control.Exception.Lens. This allows catching (and rethrowing) service specific errors returned by MediaPackageVOD.

UnprocessableEntityException

_UnprocessableEntityException :: AsError a => Getting (First ServiceError) a ServiceError

The parameters sent in the request are not valid.

ForbiddenException

_ForbiddenException :: AsError a => Getting (First ServiceError) a ServiceError

The client is not authorized to access the requested resource.

NotFoundException

_NotFoundException :: AsError a => Getting (First ServiceError) a ServiceError

The requested resource does not exist.

TooManyRequestsException

_TooManyRequestsException :: AsError a => Getting (First ServiceError) a ServiceError

The client has exceeded their resource or throttling limits.

InternalServerErrorException

_InternalServerErrorException :: AsError a => Getting (First ServiceError) a ServiceError

An unexpected error occurred.

ServiceUnavailableException

_ServiceUnavailableException :: AsError a => Getting (First ServiceError) a ServiceError

An unexpected error occurred.

Waiters

Waiters poll by repeatedly sending a request until some remote success condition configured by the Wait specification is fulfilled. The Wait specification determines how many attempts should be made, in addition to delay and retry strategies.

Operations

Some AWS operations return results that are incomplete and require subsequent requests in order to obtain the entire result set. The process of sending subsequent requests to continue where a previous request left off is called pagination. For example, the ListObjects operation of Amazon S3 returns up to 1000 objects at a time, and you must send subsequent requests with the appropriate Marker in order to retrieve the next page of results.

Operations that have an AWSPager instance can transparently perform subsequent requests, correctly setting Markers and other request facets to iterate through the entire result set of a truncated API operation. Operations which support this have an additional note in the documentation.

Many operations have the ability to filter results on the server side. See the individual operation parameters for details.

CreatePackagingGroup

data CreatePackagingGroup

A new MediaPackage VOD PackagingGroup resource configuration.

See: newCreatePackagingGroup smart constructor.

Constructors

CreatePackagingGroup' (Maybe Authorization) (Maybe EgressAccessLogs) (Maybe (HashMap Text Text)) Text

newCreatePackagingGroup

Arguments

:: Text
$sel:id:CreatePackagingGroup'

-> CreatePackagingGroup

Create a value of CreatePackagingGroup with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:authorization:CreatePackagingGroup', createPackagingGroup_authorization - Undocumented member.

$sel:egressAccessLogs:CreatePackagingGroup', createPackagingGroup_egressAccessLogs - Undocumented member.

$sel:tags:CreatePackagingGroup', createPackagingGroup_tags - Undocumented member.

$sel:id:CreatePackagingGroup', createPackagingGroup_id - The ID of the PackagingGroup.

data CreatePackagingGroupResponse

See: newCreatePackagingGroupResponse smart constructor.

Constructors

CreatePackagingGroupResponse' (Maybe Text) (Maybe Authorization) (Maybe Text) (Maybe Text) (Maybe EgressAccessLogs) (Maybe (HashMap Text Text)) Int

newCreatePackagingGroupResponse

Arguments

:: Int
$sel:httpStatus:CreatePackagingGroupResponse'

-> CreatePackagingGroupResponse

Create a value of CreatePackagingGroupResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:arn:CreatePackagingGroupResponse', createPackagingGroupResponse_arn - The ARN of the PackagingGroup.

$sel:authorization:CreatePackagingGroup', createPackagingGroupResponse_authorization - Undocumented member.

$sel:domainName:CreatePackagingGroupResponse', createPackagingGroupResponse_domainName - The fully qualified domain name for Assets in the PackagingGroup.

$sel:id:CreatePackagingGroup', createPackagingGroupResponse_id - The ID of the PackagingGroup.

$sel:egressAccessLogs:CreatePackagingGroup', createPackagingGroupResponse_egressAccessLogs - Undocumented member.

$sel:tags:CreatePackagingGroup', createPackagingGroupResponse_tags - Undocumented member.

$sel:httpStatus:CreatePackagingGroupResponse', createPackagingGroupResponse_httpStatus - The response's http status code.

ConfigureLogs

data ConfigureLogs

The option to configure log subscription.

See: newConfigureLogs smart constructor.

Constructors

ConfigureLogs' (Maybe EgressAccessLogs) Text

newConfigureLogs

Arguments

:: Text
$sel:id:ConfigureLogs'

-> ConfigureLogs

Create a value of ConfigureLogs with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:egressAccessLogs:ConfigureLogs', configureLogs_egressAccessLogs - Undocumented member.

$sel:id:ConfigureLogs', configureLogs_id - The ID of a MediaPackage VOD PackagingGroup resource.

data ConfigureLogsResponse

See: newConfigureLogsResponse smart constructor.

Constructors

ConfigureLogsResponse' (Maybe Text) (Maybe Authorization) (Maybe Text) (Maybe Text) (Maybe EgressAccessLogs) (Maybe (HashMap Text Text)) Int

newConfigureLogsResponse

Arguments

:: Int
$sel:httpStatus:ConfigureLogsResponse'

-> ConfigureLogsResponse

Create a value of ConfigureLogsResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:arn:ConfigureLogsResponse', configureLogsResponse_arn - The ARN of the PackagingGroup.

$sel:authorization:ConfigureLogsResponse', configureLogsResponse_authorization - Undocumented member.

$sel:domainName:ConfigureLogsResponse', configureLogsResponse_domainName - The fully qualified domain name for Assets in the PackagingGroup.

$sel:id:ConfigureLogs', configureLogsResponse_id - The ID of the PackagingGroup.

$sel:egressAccessLogs:ConfigureLogs', configureLogsResponse_egressAccessLogs - Undocumented member.

$sel:tags:ConfigureLogsResponse', configureLogsResponse_tags - Undocumented member.

$sel:httpStatus:ConfigureLogsResponse', configureLogsResponse_httpStatus - The response's http status code.

ListTagsForResource

data ListTagsForResource

See: newListTagsForResource smart constructor.

Constructors

ListTagsForResource' Text

newListTagsForResource

Arguments

:: Text
$sel:resourceArn:ListTagsForResource'

-> ListTagsForResource

Create a value of ListTagsForResource with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:resourceArn:ListTagsForResource', listTagsForResource_resourceArn - The Amazon Resource Name (ARN) for the resource. You can get this from the response to any request to the resource.

data ListTagsForResourceResponse

See: newListTagsForResourceResponse smart constructor.

Constructors

ListTagsForResourceResponse' (Maybe (HashMap Text Text)) Int

newListTagsForResourceResponse

Arguments

:: Int
$sel:httpStatus:ListTagsForResourceResponse'

-> ListTagsForResourceResponse

Create a value of ListTagsForResourceResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:tags:ListTagsForResourceResponse', listTagsForResourceResponse_tags - A collection of tags associated with a resource

$sel:httpStatus:ListTagsForResourceResponse', listTagsForResourceResponse_httpStatus - The response's http status code.

DescribePackagingGroup

data DescribePackagingGroup

See: newDescribePackagingGroup smart constructor.

Constructors

DescribePackagingGroup' Text

newDescribePackagingGroup

Arguments

:: Text
$sel:id:DescribePackagingGroup'

-> DescribePackagingGroup

Create a value of DescribePackagingGroup with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:id:DescribePackagingGroup', describePackagingGroup_id - The ID of a MediaPackage VOD PackagingGroup resource.

data DescribePackagingGroupResponse

See: newDescribePackagingGroupResponse smart constructor.

Constructors

DescribePackagingGroupResponse' (Maybe Text) (Maybe Authorization) (Maybe Text) (Maybe Text) (Maybe EgressAccessLogs) (Maybe (HashMap Text Text)) Int

newDescribePackagingGroupResponse

Arguments

:: Int
$sel:httpStatus:DescribePackagingGroupResponse'

-> DescribePackagingGroupResponse

Create a value of DescribePackagingGroupResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:arn:DescribePackagingGroupResponse', describePackagingGroupResponse_arn - The ARN of the PackagingGroup.

$sel:authorization:DescribePackagingGroupResponse', describePackagingGroupResponse_authorization - Undocumented member.

$sel:domainName:DescribePackagingGroupResponse', describePackagingGroupResponse_domainName - The fully qualified domain name for Assets in the PackagingGroup.

$sel:id:DescribePackagingGroup', describePackagingGroupResponse_id - The ID of the PackagingGroup.

$sel:egressAccessLogs:DescribePackagingGroupResponse', describePackagingGroupResponse_egressAccessLogs - Undocumented member.

$sel:tags:DescribePackagingGroupResponse', describePackagingGroupResponse_tags - Undocumented member.

$sel:httpStatus:DescribePackagingGroupResponse', describePackagingGroupResponse_httpStatus - The response's http status code.

DescribeAsset

data DescribeAsset

See: newDescribeAsset smart constructor.

Constructors

DescribeAsset' Text

newDescribeAsset

Arguments

:: Text
$sel:id:DescribeAsset'

-> DescribeAsset

Create a value of DescribeAsset with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:id:DescribeAsset', describeAsset_id - The ID of an MediaPackage VOD Asset resource.

data DescribeAssetResponse

See: newDescribeAssetResponse smart constructor.

Constructors

DescribeAssetResponse' (Maybe Text) (Maybe Text) (Maybe Text) (Maybe Text) (Maybe Text) (Maybe Text) (Maybe Text) (Maybe [EgressEndpoint]) (Maybe (HashMap Text Text)) Int

newDescribeAssetResponse

Arguments

:: Int
$sel:httpStatus:DescribeAssetResponse'

-> DescribeAssetResponse

Create a value of DescribeAssetResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:resourceId:DescribeAssetResponse', describeAssetResponse_resourceId - The resource ID to include in SPEKE key requests.

$sel:arn:DescribeAssetResponse', describeAssetResponse_arn - The ARN of the Asset.

$sel:createdAt:DescribeAssetResponse', describeAssetResponse_createdAt - The time the Asset was initially submitted for Ingest.

$sel:packagingGroupId:DescribeAssetResponse', describeAssetResponse_packagingGroupId - The ID of the PackagingGroup for the Asset.

$sel:sourceArn:DescribeAssetResponse', describeAssetResponse_sourceArn - ARN of the source object in S3.

$sel:sourceRoleArn:DescribeAssetResponse', describeAssetResponse_sourceRoleArn - The IAM role_arn used to access the source S3 bucket.

$sel:id:DescribeAsset', describeAssetResponse_id - The unique identifier for the Asset.

$sel:egressEndpoints:DescribeAssetResponse', describeAssetResponse_egressEndpoints - The list of egress endpoints available for the Asset.

$sel:tags:DescribeAssetResponse', describeAssetResponse_tags - Undocumented member.

$sel:httpStatus:DescribeAssetResponse', describeAssetResponse_httpStatus - The response's http status code.

DeletePackagingConfiguration

data DeletePackagingConfiguration

See: newDeletePackagingConfiguration smart constructor.

Constructors

DeletePackagingConfiguration' Text

newDeletePackagingConfiguration

Arguments

:: Text
$sel:id:DeletePackagingConfiguration'

-> DeletePackagingConfiguration

Create a value of DeletePackagingConfiguration with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:id:DeletePackagingConfiguration', deletePackagingConfiguration_id - The ID of the MediaPackage VOD PackagingConfiguration resource to delete.

data DeletePackagingConfigurationResponse

See: newDeletePackagingConfigurationResponse smart constructor.

Constructors

DeletePackagingConfigurationResponse' Int

newDeletePackagingConfigurationResponse

Arguments

:: Int
$sel:httpStatus:DeletePackagingConfigurationResponse'

-> DeletePackagingConfigurationResponse

Create a value of DeletePackagingConfigurationResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:httpStatus:DeletePackagingConfigurationResponse', deletePackagingConfigurationResponse_httpStatus - The response's http status code.

ListPackagingGroups (Paginated)

data ListPackagingGroups

See: newListPackagingGroups smart constructor.

Constructors

ListPackagingGroups' (Maybe Text) (Maybe Natural)

newListPackagingGroups :: ListPackagingGroups

Create a value of ListPackagingGroups with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:nextToken:ListPackagingGroups', listPackagingGroups_nextToken - A token used to resume pagination from the end of a previous request.

$sel:maxResults:ListPackagingGroups', listPackagingGroups_maxResults - Upper bound on number of records to return.

data ListPackagingGroupsResponse

See: newListPackagingGroupsResponse smart constructor.

Constructors

ListPackagingGroupsResponse' (Maybe [PackagingGroup]) (Maybe Text) Int

newListPackagingGroupsResponse

Arguments

:: Int
$sel:httpStatus:ListPackagingGroupsResponse'

-> ListPackagingGroupsResponse

Create a value of ListPackagingGroupsResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:packagingGroups:ListPackagingGroupsResponse', listPackagingGroupsResponse_packagingGroups - A list of MediaPackage VOD PackagingGroup resources.

$sel:nextToken:ListPackagingGroups', listPackagingGroupsResponse_nextToken - A token that can be used to resume pagination from the end of the collection.

$sel:httpStatus:ListPackagingGroupsResponse', listPackagingGroupsResponse_httpStatus - The response's http status code.

DeleteAsset

data DeleteAsset

See: newDeleteAsset smart constructor.

Constructors

DeleteAsset' Text

newDeleteAsset

Arguments

:: Text
$sel:id:DeleteAsset'

-> DeleteAsset

Create a value of DeleteAsset with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:id:DeleteAsset', deleteAsset_id - The ID of the MediaPackage VOD Asset resource to delete.

data DeleteAssetResponse

See: newDeleteAssetResponse smart constructor.

Constructors

DeleteAssetResponse' Int

newDeleteAssetResponse

Arguments

:: Int
$sel:httpStatus:DeleteAssetResponse'

-> DeleteAssetResponse

Create a value of DeleteAssetResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:httpStatus:DeleteAssetResponse', deleteAssetResponse_httpStatus - The response's http status code.

UpdatePackagingGroup

data UpdatePackagingGroup

A MediaPackage VOD PackagingGroup resource configuration.

See: newUpdatePackagingGroup smart constructor.

Constructors

UpdatePackagingGroup' (Maybe Authorization) Text

newUpdatePackagingGroup

Arguments

:: Text
$sel:id:UpdatePackagingGroup'

-> UpdatePackagingGroup

Create a value of UpdatePackagingGroup with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:authorization:UpdatePackagingGroup', updatePackagingGroup_authorization - Undocumented member.

$sel:id:UpdatePackagingGroup', updatePackagingGroup_id - The ID of a MediaPackage VOD PackagingGroup resource.

data UpdatePackagingGroupResponse

See: newUpdatePackagingGroupResponse smart constructor.

Constructors

UpdatePackagingGroupResponse' (Maybe Text) (Maybe Authorization) (Maybe Text) (Maybe Text) (Maybe EgressAccessLogs) (Maybe (HashMap Text Text)) Int

newUpdatePackagingGroupResponse

Arguments

:: Int
$sel:httpStatus:UpdatePackagingGroupResponse'

-> UpdatePackagingGroupResponse

Create a value of UpdatePackagingGroupResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:arn:UpdatePackagingGroupResponse', updatePackagingGroupResponse_arn - The ARN of the PackagingGroup.

$sel:authorization:UpdatePackagingGroup', updatePackagingGroupResponse_authorization - Undocumented member.

$sel:domainName:UpdatePackagingGroupResponse', updatePackagingGroupResponse_domainName - The fully qualified domain name for Assets in the PackagingGroup.

$sel:id:UpdatePackagingGroup', updatePackagingGroupResponse_id - The ID of the PackagingGroup.

$sel:egressAccessLogs:UpdatePackagingGroupResponse', updatePackagingGroupResponse_egressAccessLogs - Undocumented member.

$sel:tags:UpdatePackagingGroupResponse', updatePackagingGroupResponse_tags - Undocumented member.

$sel:httpStatus:UpdatePackagingGroupResponse', updatePackagingGroupResponse_httpStatus - The response's http status code.

DeletePackagingGroup

data DeletePackagingGroup

See: newDeletePackagingGroup smart constructor.

Constructors

DeletePackagingGroup' Text

newDeletePackagingGroup

Arguments

:: Text
$sel:id:DeletePackagingGroup'

-> DeletePackagingGroup

Create a value of DeletePackagingGroup with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:id:DeletePackagingGroup', deletePackagingGroup_id - The ID of the MediaPackage VOD PackagingGroup resource to delete.

data DeletePackagingGroupResponse

See: newDeletePackagingGroupResponse smart constructor.

Constructors

DeletePackagingGroupResponse' Int

newDeletePackagingGroupResponse

Arguments

:: Int
$sel:httpStatus:DeletePackagingGroupResponse'

-> DeletePackagingGroupResponse

Create a value of DeletePackagingGroupResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:httpStatus:DeletePackagingGroupResponse', deletePackagingGroupResponse_httpStatus - The response's http status code.

CreateAsset

data CreateAsset

A new MediaPackage VOD Asset configuration.

See: newCreateAsset smart constructor.

Constructors

CreateAsset' (Maybe Text) (Maybe (HashMap Text Text)) Text Text Text Text

newCreateAsset

Arguments

:: Text
$sel:sourceArn:CreateAsset'

-> Text
$sel:id:CreateAsset'

-> Text
$sel:packagingGroupId:CreateAsset'

-> Text
$sel:sourceRoleArn:CreateAsset'

-> CreateAsset

Create a value of CreateAsset with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:resourceId:CreateAsset', createAsset_resourceId - The resource ID to include in SPEKE key requests.

$sel:tags:CreateAsset', createAsset_tags - Undocumented member.

$sel:sourceArn:CreateAsset', createAsset_sourceArn - ARN of the source object in S3.

$sel:id:CreateAsset', createAsset_id - The unique identifier for the Asset.

$sel:packagingGroupId:CreateAsset', createAsset_packagingGroupId - The ID of the PackagingGroup for the Asset.

$sel:sourceRoleArn:CreateAsset', createAsset_sourceRoleArn - The IAM role ARN used to access the source S3 bucket.

data CreateAssetResponse

See: newCreateAssetResponse smart constructor.

Constructors

CreateAssetResponse' (Maybe Text) (Maybe Text) (Maybe Text) (Maybe Text) (Maybe Text) (Maybe Text) (Maybe Text) (Maybe [EgressEndpoint]) (Maybe (HashMap Text Text)) Int

newCreateAssetResponse

Arguments

:: Int
$sel:httpStatus:CreateAssetResponse'

-> CreateAssetResponse

Create a value of CreateAssetResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:resourceId:CreateAsset', createAssetResponse_resourceId - The resource ID to include in SPEKE key requests.

$sel:arn:CreateAssetResponse', createAssetResponse_arn - The ARN of the Asset.

$sel:createdAt:CreateAssetResponse', createAssetResponse_createdAt - The time the Asset was initially submitted for Ingest.

$sel:packagingGroupId:CreateAsset', createAssetResponse_packagingGroupId - The ID of the PackagingGroup for the Asset.

$sel:sourceArn:CreateAsset', createAssetResponse_sourceArn - ARN of the source object in S3.

$sel:sourceRoleArn:CreateAsset', createAssetResponse_sourceRoleArn - The IAM role_arn used to access the source S3 bucket.

$sel:id:CreateAsset', createAssetResponse_id - The unique identifier for the Asset.

$sel:egressEndpoints:CreateAssetResponse', createAssetResponse_egressEndpoints - The list of egress endpoints available for the Asset.

$sel:tags:CreateAsset', createAssetResponse_tags - Undocumented member.

$sel:httpStatus:CreateAssetResponse', createAssetResponse_httpStatus - The response's http status code.

DescribePackagingConfiguration

data DescribePackagingConfiguration

See: newDescribePackagingConfiguration smart constructor.

Constructors

DescribePackagingConfiguration' Text

newDescribePackagingConfiguration

Arguments

:: Text
$sel:id:DescribePackagingConfiguration'

-> DescribePackagingConfiguration

Create a value of DescribePackagingConfiguration with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:id:DescribePackagingConfiguration', describePackagingConfiguration_id - The ID of a MediaPackage VOD PackagingConfiguration resource.

data DescribePackagingConfigurationResponse

See: newDescribePackagingConfigurationResponse smart constructor.

Constructors

DescribePackagingConfigurationResponse' (Maybe HlsPackage) (Maybe Text) (Maybe Text) (Maybe DashPackage) (Maybe MssPackage) (Maybe Text) (Maybe CmafPackage) (Maybe (HashMap Text Text)) Int

newDescribePackagingConfigurationResponse

Arguments

:: Int
$sel:httpStatus:DescribePackagingConfigurationResponse'

-> DescribePackagingConfigurationResponse

Create a value of DescribePackagingConfigurationResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:hlsPackage:DescribePackagingConfigurationResponse', describePackagingConfigurationResponse_hlsPackage - Undocumented member.

$sel:arn:DescribePackagingConfigurationResponse', describePackagingConfigurationResponse_arn - The ARN of the PackagingConfiguration.

$sel:packagingGroupId:DescribePackagingConfigurationResponse', describePackagingConfigurationResponse_packagingGroupId - The ID of a PackagingGroup.

$sel:dashPackage:DescribePackagingConfigurationResponse', describePackagingConfigurationResponse_dashPackage - Undocumented member.

$sel:mssPackage:DescribePackagingConfigurationResponse', describePackagingConfigurationResponse_mssPackage - Undocumented member.

$sel:id:DescribePackagingConfiguration', describePackagingConfigurationResponse_id - The ID of the PackagingConfiguration.

$sel:cmafPackage:DescribePackagingConfigurationResponse', describePackagingConfigurationResponse_cmafPackage - Undocumented member.

$sel:tags:DescribePackagingConfigurationResponse', describePackagingConfigurationResponse_tags - Undocumented member.

$sel:httpStatus:DescribePackagingConfigurationResponse', describePackagingConfigurationResponse_httpStatus - The response's http status code.

TagResource

data TagResource

See: newTagResource smart constructor.

Constructors

TagResource' Text (HashMap Text Text)

newTagResource

Arguments

:: Text
$sel:resourceArn:TagResource'

-> TagResource

Create a value of TagResource with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:resourceArn:TagResource', tagResource_resourceArn - The Amazon Resource Name (ARN) for the resource. You can get this from the response to any request to the resource.

$sel:tags:TagResource', tagResource_tags - A collection of tags associated with a resource

data TagResourceResponse

See: newTagResourceResponse smart constructor.

Constructors

TagResourceResponse'

newTagResourceResponse :: TagResourceResponse

Create a value of TagResourceResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

UntagResource

data UntagResource

See: newUntagResource smart constructor.

Constructors

UntagResource' [Text] Text

newUntagResource

Arguments

:: Text
$sel:resourceArn:UntagResource'

-> UntagResource

Create a value of UntagResource with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:tagKeys:UntagResource', untagResource_tagKeys - A comma-separated list of the tag keys to remove from the resource.

$sel:resourceArn:UntagResource', untagResource_resourceArn - The Amazon Resource Name (ARN) for the resource. You can get this from the response to any request to the resource.

data UntagResourceResponse

See: newUntagResourceResponse smart constructor.

Constructors

UntagResourceResponse'

newUntagResourceResponse :: UntagResourceResponse

Create a value of UntagResourceResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

CreatePackagingConfiguration

data CreatePackagingConfiguration

A new MediaPackage VOD PackagingConfiguration resource configuration.

See: newCreatePackagingConfiguration smart constructor.

Constructors

CreatePackagingConfiguration' (Maybe HlsPackage) (Maybe DashPackage) (Maybe MssPackage) (Maybe CmafPackage) (Maybe (HashMap Text Text)) Text Text

newCreatePackagingConfiguration

Arguments

:: Text
$sel:id:CreatePackagingConfiguration'

-> Text
$sel:packagingGroupId:CreatePackagingConfiguration'

-> CreatePackagingConfiguration

Create a value of CreatePackagingConfiguration with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:hlsPackage:CreatePackagingConfiguration', createPackagingConfiguration_hlsPackage - Undocumented member.

$sel:dashPackage:CreatePackagingConfiguration', createPackagingConfiguration_dashPackage - Undocumented member.

$sel:mssPackage:CreatePackagingConfiguration', createPackagingConfiguration_mssPackage - Undocumented member.

$sel:cmafPackage:CreatePackagingConfiguration', createPackagingConfiguration_cmafPackage - Undocumented member.

$sel:tags:CreatePackagingConfiguration', createPackagingConfiguration_tags - Undocumented member.

$sel:id:CreatePackagingConfiguration', createPackagingConfiguration_id - The ID of the PackagingConfiguration.

$sel:packagingGroupId:CreatePackagingConfiguration', createPackagingConfiguration_packagingGroupId - The ID of a PackagingGroup.

data CreatePackagingConfigurationResponse

See: newCreatePackagingConfigurationResponse smart constructor.

Constructors

CreatePackagingConfigurationResponse' (Maybe HlsPackage) (Maybe Text) (Maybe Text) (Maybe DashPackage) (Maybe MssPackage) (Maybe Text) (Maybe CmafPackage) (Maybe (HashMap Text Text)) Int

newCreatePackagingConfigurationResponse

Arguments

:: Int
$sel:httpStatus:CreatePackagingConfigurationResponse'

-> CreatePackagingConfigurationResponse

Create a value of CreatePackagingConfigurationResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:hlsPackage:CreatePackagingConfiguration', createPackagingConfigurationResponse_hlsPackage - Undocumented member.

$sel:arn:CreatePackagingConfigurationResponse', createPackagingConfigurationResponse_arn - The ARN of the PackagingConfiguration.

$sel:packagingGroupId:CreatePackagingConfiguration', createPackagingConfigurationResponse_packagingGroupId - The ID of a PackagingGroup.

$sel:dashPackage:CreatePackagingConfiguration', createPackagingConfigurationResponse_dashPackage - Undocumented member.

$sel:mssPackage:CreatePackagingConfiguration', createPackagingConfigurationResponse_mssPackage - Undocumented member.

$sel:id:CreatePackagingConfiguration', createPackagingConfigurationResponse_id - The ID of the PackagingConfiguration.

$sel:cmafPackage:CreatePackagingConfiguration', createPackagingConfigurationResponse_cmafPackage - Undocumented member.

$sel:tags:CreatePackagingConfiguration', createPackagingConfigurationResponse_tags - Undocumented member.

$sel:httpStatus:CreatePackagingConfigurationResponse', createPackagingConfigurationResponse_httpStatus - The response's http status code.

ListPackagingConfigurations (Paginated)

data ListPackagingConfigurations

See: newListPackagingConfigurations smart constructor.

Constructors

ListPackagingConfigurations' (Maybe Text) (Maybe Text) (Maybe Natural)

newListPackagingConfigurations :: ListPackagingConfigurations

Create a value of ListPackagingConfigurations with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:packagingGroupId:ListPackagingConfigurations', listPackagingConfigurations_packagingGroupId - Returns MediaPackage VOD PackagingConfigurations associated with the specified PackagingGroup.

$sel:nextToken:ListPackagingConfigurations', listPackagingConfigurations_nextToken - A token used to resume pagination from the end of a previous request.

$sel:maxResults:ListPackagingConfigurations', listPackagingConfigurations_maxResults - Upper bound on number of records to return.

data ListPackagingConfigurationsResponse

See: newListPackagingConfigurationsResponse smart constructor.

Constructors

ListPackagingConfigurationsResponse' (Maybe [PackagingConfiguration]) (Maybe Text) Int

newListPackagingConfigurationsResponse

Arguments

:: Int
$sel:httpStatus:ListPackagingConfigurationsResponse'

-> ListPackagingConfigurationsResponse

Create a value of ListPackagingConfigurationsResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:packagingConfigurations:ListPackagingConfigurationsResponse', listPackagingConfigurationsResponse_packagingConfigurations - A list of MediaPackage VOD PackagingConfiguration resources.

$sel:nextToken:ListPackagingConfigurations', listPackagingConfigurationsResponse_nextToken - A token that can be used to resume pagination from the end of the collection.

$sel:httpStatus:ListPackagingConfigurationsResponse', listPackagingConfigurationsResponse_httpStatus - The response's http status code.

ListAssets (Paginated)

data ListAssets

See: newListAssets smart constructor.

Constructors

ListAssets' (Maybe Text) (Maybe Text) (Maybe Natural)

newListAssets :: ListAssets

Create a value of ListAssets with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:packagingGroupId:ListAssets', listAssets_packagingGroupId - Returns Assets associated with the specified PackagingGroup.

$sel:nextToken:ListAssets', listAssets_nextToken - A token used to resume pagination from the end of a previous request.

$sel:maxResults:ListAssets', listAssets_maxResults - Upper bound on number of records to return.

data ListAssetsResponse

See: newListAssetsResponse smart constructor.

Constructors

ListAssetsResponse' (Maybe Text) (Maybe [AssetShallow]) Int

newListAssetsResponse

Arguments

:: Int
$sel:httpStatus:ListAssetsResponse'

-> ListAssetsResponse

Create a value of ListAssetsResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:nextToken:ListAssets', listAssetsResponse_nextToken - A token that can be used to resume pagination from the end of the collection.

$sel:assets:ListAssetsResponse', listAssetsResponse_assets - A list of MediaPackage VOD Asset resources.

$sel:httpStatus:ListAssetsResponse', listAssetsResponse_httpStatus - The response's http status code.

Types

AdMarkers

newtype AdMarkers

Constructors

AdMarkers'

Fields

fromAdMarkers :: Text

Bundled Patterns

pattern AdMarkers_NONE :: AdMarkers

pattern AdMarkers_PASSTHROUGH :: AdMarkers

pattern AdMarkers_SCTE35_ENHANCED :: AdMarkers

EncryptionMethod

newtype EncryptionMethod

Constructors

EncryptionMethod'

Fields

fromEncryptionMethod :: Text

Bundled Patterns

pattern EncryptionMethod_AES_128 :: EncryptionMethod

pattern EncryptionMethod_SAMPLE_AES :: EncryptionMethod

ManifestLayout

newtype ManifestLayout

Constructors

ManifestLayout'

Fields

fromManifestLayout :: Text

Bundled Patterns

pattern ManifestLayout_COMPACT :: ManifestLayout

pattern ManifestLayout_FULL :: ManifestLayout

PeriodTriggersElement

newtype PeriodTriggersElement

Constructors

PeriodTriggersElement'

Fields

fromPeriodTriggersElement :: Text

Bundled Patterns

pattern PeriodTriggersElement_ADS :: PeriodTriggersElement

Profile

newtype Profile

Constructors

Profile'

Fields

fromProfile :: Text

Bundled Patterns

pattern Profile_HBBTV_1_5 :: Profile

pattern Profile_NONE :: Profile

SegmentTemplateFormat

newtype SegmentTemplateFormat

Constructors

SegmentTemplateFormat'

Fields

fromSegmentTemplateFormat :: Text

Bundled Patterns

pattern SegmentTemplateFormat_NUMBER_WITH_DURATION :: SegmentTemplateFormat

pattern SegmentTemplateFormat_NUMBER_WITH_TIMELINE :: SegmentTemplateFormat

pattern SegmentTemplateFormat_TIME_WITH_TIMELINE :: SegmentTemplateFormat

StreamOrder

newtype StreamOrder

Constructors

StreamOrder'

Fields

fromStreamOrder :: Text

Bundled Patterns

pattern StreamOrder_ORIGINAL :: StreamOrder

pattern StreamOrder_VIDEO_BITRATE_ASCENDING :: StreamOrder

pattern StreamOrder_VIDEO_BITRATE_DESCENDING :: StreamOrder

AssetShallow

data AssetShallow

A MediaPackage VOD Asset resource.

See: newAssetShallow smart constructor.

Constructors

AssetShallow' (Maybe Text) (Maybe Text) (Maybe Text) (Maybe Text) (Maybe Text) (Maybe Text) (Maybe Text) (Maybe (HashMap Text Text))

newAssetShallow :: AssetShallow

Create a value of AssetShallow with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:resourceId:AssetShallow', assetShallow_resourceId - The resource ID to include in SPEKE key requests.

$sel:arn:AssetShallow', assetShallow_arn - The ARN of the Asset.

$sel:createdAt:AssetShallow', assetShallow_createdAt - The time the Asset was initially submitted for Ingest.

$sel:packagingGroupId:AssetShallow', assetShallow_packagingGroupId - The ID of the PackagingGroup for the Asset.

$sel:sourceArn:AssetShallow', assetShallow_sourceArn - ARN of the source object in S3.

$sel:sourceRoleArn:AssetShallow', assetShallow_sourceRoleArn - The IAM role ARN used to access the source S3 bucket.

$sel:id:AssetShallow', assetShallow_id - The unique identifier for the Asset.

$sel:tags:AssetShallow', assetShallow_tags - Undocumented member.

Authorization

data Authorization

CDN Authorization credentials

See: newAuthorization smart constructor.

Constructors

Authorization' Text Text

newAuthorization

Arguments

:: Text
$sel:secretsRoleArn:Authorization'

-> Text
$sel:cdnIdentifierSecret:Authorization'

-> Authorization

Create a value of Authorization with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:secretsRoleArn:Authorization', authorization_secretsRoleArn - The Amazon Resource Name (ARN) for the IAM role that allows MediaPackage to communicate with AWS Secrets Manager.

$sel:cdnIdentifierSecret:Authorization', authorization_cdnIdentifierSecret - The Amazon Resource Name (ARN) for the secret in AWS Secrets Manager that is used for CDN authorization.

CmafEncryption

data CmafEncryption

A CMAF encryption configuration.

See: newCmafEncryption smart constructor.

Constructors

CmafEncryption' (Maybe Text) SpekeKeyProvider

newCmafEncryption

Arguments

:: SpekeKeyProvider
$sel:spekeKeyProvider:CmafEncryption'

-> CmafEncryption

Create a value of CmafEncryption with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:constantInitializationVector:CmafEncryption', cmafEncryption_constantInitializationVector - An optional 128-bit, 16-byte hex value represented by a 32-character string, used in conjunction with the key for encrypting blocks. If you don't specify a value, then MediaPackage creates the constant initialization vector (IV).

$sel:spekeKeyProvider:CmafEncryption', cmafEncryption_spekeKeyProvider - Undocumented member.

CmafPackage

data CmafPackage

A CMAF packaging configuration.

See: newCmafPackage smart constructor.

Constructors

CmafPackage' (Maybe Bool) (Maybe Int) (Maybe CmafEncryption) [HlsManifest]

newCmafPackage :: CmafPackage

Create a value of CmafPackage with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:includeEncoderConfigurationInSegments:CmafPackage', cmafPackage_includeEncoderConfigurationInSegments - When includeEncoderConfigurationInSegments is set to true, MediaPackage places your encoder's Sequence Parameter Set (SPS), Picture Parameter Set (PPS), and Video Parameter Set (VPS) metadata in every video segment instead of in the init fragment. This lets you use different SPS/PPS/VPS settings for your assets during content playback.

$sel:segmentDurationSeconds:CmafPackage', cmafPackage_segmentDurationSeconds - Duration (in seconds) of each fragment. Actual fragments will be rounded to the nearest multiple of the source fragment duration.

$sel:encryption:CmafPackage', cmafPackage_encryption - Undocumented member.

$sel:hlsManifests:CmafPackage', cmafPackage_hlsManifests - A list of HLS manifest configurations.

DashEncryption

data DashEncryption

A Dynamic Adaptive Streaming over HTTP (DASH) encryption configuration.

See: newDashEncryption smart constructor.

Constructors

DashEncryption' SpekeKeyProvider

newDashEncryption

Arguments

:: SpekeKeyProvider
$sel:spekeKeyProvider:DashEncryption'

-> DashEncryption

Create a value of DashEncryption with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:spekeKeyProvider:DashEncryption', dashEncryption_spekeKeyProvider - Undocumented member.

DashManifest

data DashManifest

A DASH manifest configuration.

See: newDashManifest smart constructor.

Constructors

DashManifest' (Maybe Int) (Maybe Text) (Maybe Profile) (Maybe StreamSelection) (Maybe ManifestLayout)

newDashManifest :: DashManifest

Create a value of DashManifest with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:minBufferTimeSeconds:DashManifest', dashManifest_minBufferTimeSeconds - Minimum duration (in seconds) that a player will buffer media before starting the presentation.

$sel:manifestName:DashManifest', dashManifest_manifestName - An optional string to include in the name of the manifest.

$sel:profile:DashManifest', dashManifest_profile - The Dynamic Adaptive Streaming over HTTP (DASH) profile type. When set to "HBBTV_1_5", HbbTV 1.5 compliant output is enabled.

$sel:streamSelection:DashManifest', dashManifest_streamSelection - Undocumented member.

$sel:manifestLayout:DashManifest', dashManifest_manifestLayout - Determines the position of some tags in the Media Presentation Description (MPD). When set to FULL, elements like SegmentTemplate and ContentProtection are included in each Representation. When set to COMPACT, duplicate elements are combined and presented at the AdaptationSet level.

DashPackage

data DashPackage

A Dynamic Adaptive Streaming over HTTP (DASH) packaging configuration.

See: newDashPackage smart constructor.

Constructors

DashPackage' (Maybe Bool) (Maybe SegmentTemplateFormat) (Maybe Int) (Maybe DashEncryption) (Maybe [PeriodTriggersElement]) [DashManifest]

newDashPackage :: DashPackage

Create a value of DashPackage with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:includeEncoderConfigurationInSegments:DashPackage', dashPackage_includeEncoderConfigurationInSegments - When includeEncoderConfigurationInSegments is set to true, MediaPackage places your encoder's Sequence Parameter Set (SPS), Picture Parameter Set (PPS), and Video Parameter Set (VPS) metadata in every video segment instead of in the init fragment. This lets you use different SPS/PPS/VPS settings for your assets during content playback.

$sel:segmentTemplateFormat:DashPackage', dashPackage_segmentTemplateFormat - Determines the type of SegmentTemplate included in the Media Presentation Description (MPD). When set to NUMBER_WITH_TIMELINE, a full timeline is presented in each SegmentTemplate, with $Number$ media URLs. When set to TIME_WITH_TIMELINE, a full timeline is presented in each SegmentTemplate, with $Time$ media URLs. When set to NUMBER_WITH_DURATION, only a duration is included in each SegmentTemplate, with $Number$ media URLs.

$sel:segmentDurationSeconds:DashPackage', dashPackage_segmentDurationSeconds - Duration (in seconds) of each segment. Actual segments will be rounded to the nearest multiple of the source segment duration.

$sel:encryption:DashPackage', dashPackage_encryption - Undocumented member.

$sel:periodTriggers:DashPackage', dashPackage_periodTriggers - A list of triggers that controls when the outgoing Dynamic Adaptive Streaming over HTTP (DASH) Media Presentation Description (MPD) will be partitioned into multiple periods. If empty, the content will not be partitioned into more than one period. If the list contains "ADS", new periods will be created where the Asset contains SCTE-35 ad markers.

$sel:dashManifests:DashPackage', dashPackage_dashManifests - A list of DASH manifest configurations.

EgressAccessLogs

data EgressAccessLogs

Configure egress access logging.

See: newEgressAccessLogs smart constructor.

Constructors

EgressAccessLogs' (Maybe Text)

newEgressAccessLogs :: EgressAccessLogs

Create a value of EgressAccessLogs with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:logGroupName:EgressAccessLogs', egressAccessLogs_logGroupName - Customize the log group name.

EgressEndpoint

data EgressEndpoint

The endpoint URL used to access an Asset using one PackagingConfiguration.

See: newEgressEndpoint smart constructor.

Constructors

EgressEndpoint' (Maybe Text) (Maybe Text) (Maybe Text)

newEgressEndpoint :: EgressEndpoint

Create a value of EgressEndpoint with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:status:EgressEndpoint', egressEndpoint_status - The current processing status of the asset used for the packaging configuration. The status can be either QUEUED, PROCESSING, PLAYABLE, or FAILED. Status information won't be available for most assets ingested before 2021-09-30.

$sel:url:EgressEndpoint', egressEndpoint_url - The URL of the parent manifest for the repackaged Asset.

$sel:packagingConfigurationId:EgressEndpoint', egressEndpoint_packagingConfigurationId - The ID of the PackagingConfiguration being applied to the Asset.

HlsEncryption

data HlsEncryption

An HTTP Live Streaming (HLS) encryption configuration.

See: newHlsEncryption smart constructor.

Constructors

HlsEncryption' (Maybe EncryptionMethod) (Maybe Text) SpekeKeyProvider

newHlsEncryption

Arguments

:: SpekeKeyProvider
$sel:spekeKeyProvider:HlsEncryption'

-> HlsEncryption

Create a value of HlsEncryption with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:encryptionMethod:HlsEncryption', hlsEncryption_encryptionMethod - The encryption method to use.

$sel:constantInitializationVector:HlsEncryption', hlsEncryption_constantInitializationVector - A constant initialization vector for encryption (optional). When not specified the initialization vector will be periodically rotated.

$sel:spekeKeyProvider:HlsEncryption', hlsEncryption_spekeKeyProvider - Undocumented member.

HlsManifest

data HlsManifest

An HTTP Live Streaming (HLS) manifest configuration.

See: newHlsManifest smart constructor.

Constructors

HlsManifest' (Maybe Text) (Maybe Int) (Maybe StreamSelection) (Maybe AdMarkers) (Maybe Bool) (Maybe Bool)

newHlsManifest :: HlsManifest

Create a value of HlsManifest with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:manifestName:HlsManifest', hlsManifest_manifestName - An optional string to include in the name of the manifest.

$sel:programDateTimeIntervalSeconds:HlsManifest', hlsManifest_programDateTimeIntervalSeconds - The interval (in seconds) between each EXT-X-PROGRAM-DATE-TIME tag inserted into manifests. Additionally, when an interval is specified ID3Timed Metadata messages will be generated every 5 seconds using the ingest time of the content. If the interval is not specified, or set to 0, then no EXT-X-PROGRAM-DATE-TIME tags will be inserted into manifests and no ID3Timed Metadata messages will be generated. Note that irrespective of this parameter, if any ID3 Timed Metadata is found in HTTP Live Streaming (HLS) input, it will be passed through to HLS output.

$sel:streamSelection:HlsManifest', hlsManifest_streamSelection - Undocumented member.

$sel:adMarkers:HlsManifest', hlsManifest_adMarkers - This setting controls how ad markers are included in the packaged OriginEndpoint. "NONE" will omit all SCTE-35 ad markers from the output. "PASSTHROUGH" causes the manifest to contain a copy of the SCTE-35 ad markers (comments) taken directly from the input HTTP Live Streaming (HLS) manifest. "SCTE35_ENHANCED" generates ad markers and blackout tags based on SCTE-35 messages in the input source.

$sel:includeIframeOnlyStream:HlsManifest', hlsManifest_includeIframeOnlyStream - When enabled, an I-Frame only stream will be included in the output.

$sel:repeatExtXKey:HlsManifest', hlsManifest_repeatExtXKey - When enabled, the EXT-X-KEY tag will be repeated in output manifests.

HlsPackage

data HlsPackage

An HTTP Live Streaming (HLS) packaging configuration.

See: newHlsPackage smart constructor.

Constructors

HlsPackage' (Maybe Bool) (Maybe Bool) (Maybe Int) (Maybe HlsEncryption) [HlsManifest]

newHlsPackage :: HlsPackage

Create a value of HlsPackage with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:useAudioRenditionGroup:HlsPackage', hlsPackage_useAudioRenditionGroup - When enabled, audio streams will be placed in rendition groups in the output.

$sel:includeDvbSubtitles:HlsPackage', hlsPackage_includeDvbSubtitles - When enabled, MediaPackage passes through digital video broadcasting (DVB) subtitles into the output.

$sel:segmentDurationSeconds:HlsPackage', hlsPackage_segmentDurationSeconds - Duration (in seconds) of each fragment. Actual fragments will be rounded to the nearest multiple of the source fragment duration.

$sel:encryption:HlsPackage', hlsPackage_encryption - Undocumented member.

$sel:hlsManifests:HlsPackage', hlsPackage_hlsManifests - A list of HLS manifest configurations.

MssEncryption

data MssEncryption

A Microsoft Smooth Streaming (MSS) encryption configuration.

See: newMssEncryption smart constructor.

Constructors

MssEncryption' SpekeKeyProvider

newMssEncryption

Arguments

:: SpekeKeyProvider
$sel:spekeKeyProvider:MssEncryption'

-> MssEncryption

Create a value of MssEncryption with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:spekeKeyProvider:MssEncryption', mssEncryption_spekeKeyProvider - Undocumented member.

MssManifest

data MssManifest

A Microsoft Smooth Streaming (MSS) manifest configuration.

See: newMssManifest smart constructor.

Constructors

MssManifest' (Maybe Text) (Maybe StreamSelection)

newMssManifest :: MssManifest

Create a value of MssManifest with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:manifestName:MssManifest', mssManifest_manifestName - An optional string to include in the name of the manifest.

$sel:streamSelection:MssManifest', mssManifest_streamSelection - Undocumented member.

MssPackage

data MssPackage

A Microsoft Smooth Streaming (MSS) PackagingConfiguration.

See: newMssPackage smart constructor.

Constructors

MssPackage' (Maybe Int) (Maybe MssEncryption) [MssManifest]

newMssPackage :: MssPackage

Create a value of MssPackage with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:segmentDurationSeconds:MssPackage', mssPackage_segmentDurationSeconds - The duration (in seconds) of each segment.

$sel:encryption:MssPackage', mssPackage_encryption - Undocumented member.

$sel:mssManifests:MssPackage', mssPackage_mssManifests - A list of MSS manifest configurations.

PackagingConfiguration

data PackagingConfiguration

A MediaPackage VOD PackagingConfiguration resource.

See: newPackagingConfiguration smart constructor.

Constructors

PackagingConfiguration' (Maybe HlsPackage) (Maybe Text) (Maybe Text) (Maybe DashPackage) (Maybe MssPackage) (Maybe Text) (Maybe CmafPackage) (Maybe (HashMap Text Text))

newPackagingConfiguration :: PackagingConfiguration

Create a value of PackagingConfiguration with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:hlsPackage:PackagingConfiguration', packagingConfiguration_hlsPackage - Undocumented member.

$sel:arn:PackagingConfiguration', packagingConfiguration_arn - The ARN of the PackagingConfiguration.

$sel:packagingGroupId:PackagingConfiguration', packagingConfiguration_packagingGroupId - The ID of a PackagingGroup.

$sel:dashPackage:PackagingConfiguration', packagingConfiguration_dashPackage - Undocumented member.

$sel:mssPackage:PackagingConfiguration', packagingConfiguration_mssPackage - Undocumented member.

$sel:id:PackagingConfiguration', packagingConfiguration_id - The ID of the PackagingConfiguration.

$sel:cmafPackage:PackagingConfiguration', packagingConfiguration_cmafPackage - Undocumented member.

$sel:tags:PackagingConfiguration', packagingConfiguration_tags - Undocumented member.

PackagingGroup

data PackagingGroup

A MediaPackage VOD PackagingGroup resource.

See: newPackagingGroup smart constructor.

Constructors

PackagingGroup' (Maybe Text) (Maybe Authorization) (Maybe Text) (Maybe Text) (Maybe EgressAccessLogs) (Maybe (HashMap Text Text))

newPackagingGroup :: PackagingGroup

Create a value of PackagingGroup with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:arn:PackagingGroup', packagingGroup_arn - The ARN of the PackagingGroup.

$sel:authorization:PackagingGroup', packagingGroup_authorization - Undocumented member.

$sel:domainName:PackagingGroup', packagingGroup_domainName - The fully qualified domain name for Assets in the PackagingGroup.

$sel:id:PackagingGroup', packagingGroup_id - The ID of the PackagingGroup.

$sel:egressAccessLogs:PackagingGroup', packagingGroup_egressAccessLogs - Undocumented member.

$sel:tags:PackagingGroup', packagingGroup_tags - Undocumented member.

SpekeKeyProvider

data SpekeKeyProvider

A configuration for accessing an external Secure Packager and Encoder Key Exchange (SPEKE) service that will provide encryption keys.

See: newSpekeKeyProvider smart constructor.

Constructors

SpekeKeyProvider' [Text] Text Text

newSpekeKeyProvider

Arguments

:: Text
$sel:url:SpekeKeyProvider'

-> Text
$sel:roleArn:SpekeKeyProvider'

-> SpekeKeyProvider

Create a value of SpekeKeyProvider with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:systemIds:SpekeKeyProvider', spekeKeyProvider_systemIds - The system IDs to include in key requests.

$sel:url:SpekeKeyProvider', spekeKeyProvider_url - The URL of the external key provider service.

$sel:roleArn:SpekeKeyProvider', spekeKeyProvider_roleArn - An Amazon Resource Name (ARN) of an IAM role that AWS Elemental MediaPackage will assume when accessing the key provider service.

StreamSelection

data StreamSelection

A StreamSelection configuration.

See: newStreamSelection smart constructor.

Constructors

StreamSelection' (Maybe StreamOrder) (Maybe Int) (Maybe Int)

newStreamSelection :: StreamSelection

Create a value of StreamSelection with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:streamOrder:StreamSelection', streamSelection_streamOrder - A directive that determines the order of streams in the output.

$sel:minVideoBitsPerSecond:StreamSelection', streamSelection_minVideoBitsPerSecond - The minimum video bitrate (bps) to include in output.

$sel:maxVideoBitsPerSecond:StreamSelection', streamSelection_maxVideoBitsPerSecond - The maximum video bitrate (bps) to include in output.