Amazonka.MediaPackageVOD.CreateAsset

Description

Creates a new MediaPackage VOD Asset resource.

Creating a Request

data CreateAsset

A new MediaPackage VOD Asset configuration.

See: newCreateAsset smart constructor.

Constructors

CreateAsset'

Fields

resourceId :: Maybe Text

The resource ID to include in SPEKE key requests.

tags :: Maybe (HashMap Text Text)

sourceArn :: Text

ARN of the source object in S3.

id :: Text

The unique identifier for the Asset.

packagingGroupId :: Text

The ID of the PackagingGroup for the Asset.

sourceRoleArn :: Text

The IAM role ARN used to access the source S3 bucket.

newCreateAsset

Arguments

:: Text
$sel:sourceArn:CreateAsset'

-> Text
$sel:id:CreateAsset'

-> Text
$sel:packagingGroupId:CreateAsset'

-> Text
$sel:sourceRoleArn:CreateAsset'

-> CreateAsset

Create a value of CreateAsset with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:resourceId:CreateAsset', createAsset_resourceId - The resource ID to include in SPEKE key requests.

$sel:tags:CreateAsset', createAsset_tags - Undocumented member.

$sel:sourceArn:CreateAsset', createAsset_sourceArn - ARN of the source object in S3.

$sel:id:CreateAsset', createAsset_id - The unique identifier for the Asset.

$sel:packagingGroupId:CreateAsset', createAsset_packagingGroupId - The ID of the PackagingGroup for the Asset.

$sel:sourceRoleArn:CreateAsset', createAsset_sourceRoleArn - The IAM role ARN used to access the source S3 bucket.

Request Lenses

createAsset_resourceId :: Lens' CreateAsset (Maybe Text)

The resource ID to include in SPEKE key requests.

createAsset_tags :: Lens' CreateAsset (Maybe (HashMap Text Text))

Undocumented member.

createAsset_sourceArn :: Lens' CreateAsset Text

ARN of the source object in S3.

createAsset_id :: Lens' CreateAsset Text

The unique identifier for the Asset.

createAsset_packagingGroupId :: Lens' CreateAsset Text

The ID of the PackagingGroup for the Asset.

createAsset_sourceRoleArn :: Lens' CreateAsset Text

The IAM role ARN used to access the source S3 bucket.

Destructuring the Response

data CreateAssetResponse

See: newCreateAssetResponse smart constructor.

Constructors

CreateAssetResponse'

Fields

resourceId :: Maybe Text

The resource ID to include in SPEKE key requests.

arn :: Maybe Text

The ARN of the Asset.

createdAt :: Maybe Text

The time the Asset was initially submitted for Ingest.

packagingGroupId :: Maybe Text

The ID of the PackagingGroup for the Asset.

sourceArn :: Maybe Text

ARN of the source object in S3.

sourceRoleArn :: Maybe Text

The IAM role_arn used to access the source S3 bucket.

id :: Maybe Text

The unique identifier for the Asset.

egressEndpoints :: Maybe [EgressEndpoint]

The list of egress endpoints available for the Asset.

tags :: Maybe (HashMap Text Text)

httpStatus :: Int

The response's http status code.

newCreateAssetResponse

Arguments

:: Int
$sel:httpStatus:CreateAssetResponse'

-> CreateAssetResponse

Create a value of CreateAssetResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:resourceId:CreateAsset', createAssetResponse_resourceId - The resource ID to include in SPEKE key requests.

$sel:arn:CreateAssetResponse', createAssetResponse_arn - The ARN of the Asset.

$sel:createdAt:CreateAssetResponse', createAssetResponse_createdAt - The time the Asset was initially submitted for Ingest.

$sel:packagingGroupId:CreateAsset', createAssetResponse_packagingGroupId - The ID of the PackagingGroup for the Asset.

$sel:sourceArn:CreateAsset', createAssetResponse_sourceArn - ARN of the source object in S3.

$sel:sourceRoleArn:CreateAsset', createAssetResponse_sourceRoleArn - The IAM role_arn used to access the source S3 bucket.

$sel:id:CreateAsset', createAssetResponse_id - The unique identifier for the Asset.

$sel:egressEndpoints:CreateAssetResponse', createAssetResponse_egressEndpoints - The list of egress endpoints available for the Asset.

$sel:tags:CreateAsset', createAssetResponse_tags - Undocumented member.

$sel:httpStatus:CreateAssetResponse', createAssetResponse_httpStatus - The response's http status code.

Response Lenses

createAssetResponse_resourceId :: Lens' CreateAssetResponse (Maybe Text)

The resource ID to include in SPEKE key requests.

createAssetResponse_arn :: Lens' CreateAssetResponse (Maybe Text)

The ARN of the Asset.

createAssetResponse_createdAt :: Lens' CreateAssetResponse (Maybe Text)

The time the Asset was initially submitted for Ingest.

createAssetResponse_packagingGroupId :: Lens' CreateAssetResponse (Maybe Text)

The ID of the PackagingGroup for the Asset.

createAssetResponse_sourceArn :: Lens' CreateAssetResponse (Maybe Text)

ARN of the source object in S3.

createAssetResponse_sourceRoleArn :: Lens' CreateAssetResponse (Maybe Text)

The IAM role_arn used to access the source S3 bucket.

createAssetResponse_id :: Lens' CreateAssetResponse (Maybe Text)

The unique identifier for the Asset.

createAssetResponse_egressEndpoints :: Lens' CreateAssetResponse (Maybe [EgressEndpoint])

The list of egress endpoints available for the Asset.

createAssetResponse_tags :: Lens' CreateAssetResponse (Maybe (HashMap Text Text))

Undocumented member.

createAssetResponse_httpStatus :: Lens' CreateAssetResponse Int

The response's http status code.