Amazonka.MediaPackageVOD.CreatePackagingConfiguration

Description

Creates a new MediaPackage VOD PackagingConfiguration resource.

Creating a Request

data CreatePackagingConfiguration

A new MediaPackage VOD PackagingConfiguration resource configuration.

See: newCreatePackagingConfiguration smart constructor.

Constructors

CreatePackagingConfiguration'

Fields

hlsPackage :: Maybe HlsPackage

dashPackage :: Maybe DashPackage

mssPackage :: Maybe MssPackage

cmafPackage :: Maybe CmafPackage

tags :: Maybe (HashMap Text Text)

id :: Text

The ID of the PackagingConfiguration.

packagingGroupId :: Text

The ID of a PackagingGroup.

newCreatePackagingConfiguration

Arguments

:: Text
$sel:id:CreatePackagingConfiguration'

-> Text
$sel:packagingGroupId:CreatePackagingConfiguration'

-> CreatePackagingConfiguration

Create a value of CreatePackagingConfiguration with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:hlsPackage:CreatePackagingConfiguration', createPackagingConfiguration_hlsPackage - Undocumented member.

$sel:dashPackage:CreatePackagingConfiguration', createPackagingConfiguration_dashPackage - Undocumented member.

$sel:mssPackage:CreatePackagingConfiguration', createPackagingConfiguration_mssPackage - Undocumented member.

$sel:cmafPackage:CreatePackagingConfiguration', createPackagingConfiguration_cmafPackage - Undocumented member.

$sel:tags:CreatePackagingConfiguration', createPackagingConfiguration_tags - Undocumented member.

$sel:id:CreatePackagingConfiguration', createPackagingConfiguration_id - The ID of the PackagingConfiguration.

$sel:packagingGroupId:CreatePackagingConfiguration', createPackagingConfiguration_packagingGroupId - The ID of a PackagingGroup.

Request Lenses

createPackagingConfiguration_hlsPackage :: Lens' CreatePackagingConfiguration (Maybe HlsPackage)

Undocumented member.

createPackagingConfiguration_dashPackage :: Lens' CreatePackagingConfiguration (Maybe DashPackage)

Undocumented member.

createPackagingConfiguration_mssPackage :: Lens' CreatePackagingConfiguration (Maybe MssPackage)

Undocumented member.

createPackagingConfiguration_cmafPackage :: Lens' CreatePackagingConfiguration (Maybe CmafPackage)

Undocumented member.

createPackagingConfiguration_tags :: Lens' CreatePackagingConfiguration (Maybe (HashMap Text Text))

Undocumented member.

createPackagingConfiguration_id :: Lens' CreatePackagingConfiguration Text

The ID of the PackagingConfiguration.

createPackagingConfiguration_packagingGroupId :: Lens' CreatePackagingConfiguration Text

The ID of a PackagingGroup.

Destructuring the Response

data CreatePackagingConfigurationResponse

See: newCreatePackagingConfigurationResponse smart constructor.

Constructors

CreatePackagingConfigurationResponse'

Fields

hlsPackage :: Maybe HlsPackage

arn :: Maybe Text

The ARN of the PackagingConfiguration.

packagingGroupId :: Maybe Text

The ID of a PackagingGroup.

dashPackage :: Maybe DashPackage

mssPackage :: Maybe MssPackage

id :: Maybe Text

The ID of the PackagingConfiguration.

cmafPackage :: Maybe CmafPackage

tags :: Maybe (HashMap Text Text)

httpStatus :: Int

The response's http status code.

newCreatePackagingConfigurationResponse

Arguments

:: Int
$sel:httpStatus:CreatePackagingConfigurationResponse'

-> CreatePackagingConfigurationResponse

Create a value of CreatePackagingConfigurationResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:hlsPackage:CreatePackagingConfiguration', createPackagingConfigurationResponse_hlsPackage - Undocumented member.

$sel:arn:CreatePackagingConfigurationResponse', createPackagingConfigurationResponse_arn - The ARN of the PackagingConfiguration.

$sel:packagingGroupId:CreatePackagingConfiguration', createPackagingConfigurationResponse_packagingGroupId - The ID of a PackagingGroup.

$sel:dashPackage:CreatePackagingConfiguration', createPackagingConfigurationResponse_dashPackage - Undocumented member.

$sel:mssPackage:CreatePackagingConfiguration', createPackagingConfigurationResponse_mssPackage - Undocumented member.

$sel:id:CreatePackagingConfiguration', createPackagingConfigurationResponse_id - The ID of the PackagingConfiguration.

$sel:cmafPackage:CreatePackagingConfiguration', createPackagingConfigurationResponse_cmafPackage - Undocumented member.

$sel:tags:CreatePackagingConfiguration', createPackagingConfigurationResponse_tags - Undocumented member.

$sel:httpStatus:CreatePackagingConfigurationResponse', createPackagingConfigurationResponse_httpStatus - The response's http status code.

Response Lenses

createPackagingConfigurationResponse_hlsPackage :: Lens' CreatePackagingConfigurationResponse (Maybe HlsPackage)

Undocumented member.

createPackagingConfigurationResponse_arn :: Lens' CreatePackagingConfigurationResponse (Maybe Text)

The ARN of the PackagingConfiguration.

createPackagingConfigurationResponse_packagingGroupId :: Lens' CreatePackagingConfigurationResponse (Maybe Text)

The ID of a PackagingGroup.

createPackagingConfigurationResponse_dashPackage :: Lens' CreatePackagingConfigurationResponse (Maybe DashPackage)

Undocumented member.

createPackagingConfigurationResponse_mssPackage :: Lens' CreatePackagingConfigurationResponse (Maybe MssPackage)

Undocumented member.

createPackagingConfigurationResponse_id :: Lens' CreatePackagingConfigurationResponse (Maybe Text)

The ID of the PackagingConfiguration.

createPackagingConfigurationResponse_cmafPackage :: Lens' CreatePackagingConfigurationResponse (Maybe CmafPackage)

Undocumented member.

createPackagingConfigurationResponse_tags :: Lens' CreatePackagingConfigurationResponse (Maybe (HashMap Text Text))

Undocumented member.

createPackagingConfigurationResponse_httpStatus :: Lens' CreatePackagingConfigurationResponse Int

The response's http status code.