Amazonka.MediaPackageVOD.DeleteAsset

Description

Deletes an existing MediaPackage VOD Asset resource.

Creating a Request

data DeleteAsset

See: newDeleteAsset smart constructor.

Constructors

DeleteAsset'

Fields

id :: Text

The ID of the MediaPackage VOD Asset resource to delete.

newDeleteAsset

Arguments

:: Text
$sel:id:DeleteAsset'

-> DeleteAsset

Create a value of DeleteAsset with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:id:DeleteAsset', deleteAsset_id - The ID of the MediaPackage VOD Asset resource to delete.

Request Lenses

deleteAsset_id :: Lens' DeleteAsset Text

The ID of the MediaPackage VOD Asset resource to delete.

Destructuring the Response

data DeleteAssetResponse

See: newDeleteAssetResponse smart constructor.

Constructors

DeleteAssetResponse'

Fields

httpStatus :: Int

The response's http status code.

newDeleteAssetResponse

Arguments

:: Int
$sel:httpStatus:DeleteAssetResponse'

-> DeleteAssetResponse

Create a value of DeleteAssetResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:httpStatus:DeleteAssetResponse', deleteAssetResponse_httpStatus - The response's http status code.

Response Lenses

deleteAssetResponse_httpStatus :: Lens' DeleteAssetResponse Int

The response's http status code.